Network.AWS.CloudHSMV2.Lens

Description

Operations

DescribeClusters

describeClusters_filters :: Lens' DescribeClusters (Maybe (HashMap Text [Text]))

One or more filters to limit the items returned in the response.

Use the clusterIds filter to return only the specified clusters. Specify clusters by their cluster identifier (ID).

Use the vpcIds filter to return only the clusters in the specified virtual private clouds (VPCs). Specify VPCs by their VPC identifier (ID).

Use the states filter to return only clusters that match the specified state.

describeClusters_nextToken :: Lens' DescribeClusters (Maybe Text)

The NextToken value that you received in the previous response. Use this value to get more clusters.

describeClusters_maxResults :: Lens' DescribeClusters (Maybe Natural)

The maximum number of clusters to return in the response. When there are more clusters than the number you specify, the response contains a NextToken value.

describeClustersResponse_nextToken :: Lens' DescribeClustersResponse (Maybe Text)

An opaque string that indicates that the response contains only a subset of clusters. Use this value in a subsequent DescribeClusters request to get more clusters.

describeClustersResponse_clusters :: Lens' DescribeClustersResponse (Maybe [Cluster])

A list of clusters.

describeClustersResponse_httpStatus :: Lens' DescribeClustersResponse Int

The response's http status code.

DeleteBackup

deleteBackup_backupId :: Lens' DeleteBackup Text

The ID of the backup to be deleted. To find the ID of a backup, use the DescribeBackups operation.

deleteBackupResponse_backup :: Lens' DeleteBackupResponse (Maybe Backup)

Information on the Backup object deleted.

deleteBackupResponse_httpStatus :: Lens' DeleteBackupResponse Int

The response's http status code.

InitializeCluster

initializeCluster_clusterId :: Lens' InitializeCluster Text

The identifier (ID) of the cluster that you are claiming. To find the cluster ID, use DescribeClusters.

initializeCluster_signedCert :: Lens' InitializeCluster Text

The cluster certificate issued (signed) by your issuing certificate authority (CA). The certificate must be in PEM format and can contain a maximum of 5000 characters.

initializeCluster_trustAnchor :: Lens' InitializeCluster Text

The issuing certificate of the issuing certificate authority (CA) that issued (signed) the cluster certificate. You must use a self-signed certificate. The certificate used to sign the HSM CSR must be directly available, and thus must be the root certificate. The certificate must be in PEM format and can contain a maximum of 5000 characters.

initializeClusterResponse_stateMessage :: Lens' InitializeClusterResponse (Maybe Text)

A description of the cluster's state.

initializeClusterResponse_state :: Lens' InitializeClusterResponse (Maybe ClusterState)

The cluster's state.

initializeClusterResponse_httpStatus :: Lens' InitializeClusterResponse Int

The response's http status code.

CreateHsm

createHsm_ipAddress :: Lens' CreateHsm (Maybe Text)

The HSM's IP address. If you specify an IP address, use an available address from the subnet that maps to the Availability Zone where you are creating the HSM. If you don't specify an IP address, one is chosen for you from that subnet.

createHsm_clusterId :: Lens' CreateHsm Text

The identifier (ID) of the HSM's cluster. To find the cluster ID, use DescribeClusters.

createHsm_availabilityZone :: Lens' CreateHsm Text

The Availability Zone where you are creating the HSM. To find the cluster's Availability Zones, use DescribeClusters.

createHsmResponse_hsm :: Lens' CreateHsmResponse (Maybe Hsm)

Information about the HSM that was created.

createHsmResponse_httpStatus :: Lens' CreateHsmResponse Int

The response's http status code.

DescribeBackups

describeBackups_sortAscending :: Lens' DescribeBackups (Maybe Bool)

Designates whether or not to sort the return backups by ascending chronological order of generation.

describeBackups_filters :: Lens' DescribeBackups (Maybe (HashMap Text [Text]))

One or more filters to limit the items returned in the response.

Use the backupIds filter to return only the specified backups. Specify backups by their backup identifier (ID).

Use the sourceBackupIds filter to return only the backups created from a source backup. The sourceBackupID of a source backup is returned by the CopyBackupToRegion operation.

Use the clusterIds filter to return only the backups for the specified clusters. Specify clusters by their cluster identifier (ID).

Use the states filter to return only backups that match the specified state.

Use the neverExpires filter to return backups filtered by the value in the neverExpires parameter. True returns all backups exempt from the backup retention policy. False returns all backups with a backup retention policy defined at the cluster.

describeBackups_nextToken :: Lens' DescribeBackups (Maybe Text)

The NextToken value that you received in the previous response. Use this value to get more backups.

describeBackups_maxResults :: Lens' DescribeBackups (Maybe Natural)

The maximum number of backups to return in the response. When there are more backups than the number you specify, the response contains a NextToken value.

describeBackupsResponse_backups :: Lens' DescribeBackupsResponse (Maybe [Backup])

A list of backups.

describeBackupsResponse_nextToken :: Lens' DescribeBackupsResponse (Maybe Text)

An opaque string that indicates that the response contains only a subset of backups. Use this value in a subsequent DescribeBackups request to get more backups.

describeBackupsResponse_httpStatus :: Lens' DescribeBackupsResponse Int

The response's http status code.

CopyBackupToRegion

copyBackupToRegion_tagList :: Lens' CopyBackupToRegion (Maybe [Tag])

Tags to apply to the destination backup during creation. If you specify tags, only these tags will be applied to the destination backup. If you do not specify tags, the service copies tags from the source backup to the destination backup.

copyBackupToRegion_destinationRegion :: Lens' CopyBackupToRegion Text

The AWS region that will contain your copied CloudHSM cluster backup.

copyBackupToRegion_backupId :: Lens' CopyBackupToRegion Text

The ID of the backup that will be copied to the destination region.

copyBackupToRegionResponse_destinationBackup :: Lens' CopyBackupToRegionResponse (Maybe DestinationBackup)

Information on the backup that will be copied to the destination region, including CreateTimestamp, SourceBackup, SourceCluster, and Source Region. CreateTimestamp of the destination backup will be the same as that of the source backup.

You will need to use the sourceBackupID returned in this operation to use the DescribeBackups operation on the backup that will be copied to the destination region.

copyBackupToRegionResponse_httpStatus :: Lens' CopyBackupToRegionResponse Int

The response's http status code.

DeleteCluster

deleteCluster_clusterId :: Lens' DeleteCluster Text

The identifier (ID) of the cluster that you are deleting. To find the cluster ID, use DescribeClusters.

deleteClusterResponse_cluster :: Lens' DeleteClusterResponse (Maybe Cluster)

Information about the cluster that was deleted.

deleteClusterResponse_httpStatus :: Lens' DeleteClusterResponse Int

The response's http status code.

CreateCluster

createCluster_backupRetentionPolicy :: Lens' CreateCluster (Maybe BackupRetentionPolicy)

A policy that defines how the service retains backups.

createCluster_tagList :: Lens' CreateCluster (Maybe [Tag])

Tags to apply to the CloudHSM cluster during creation.

createCluster_sourceBackupId :: Lens' CreateCluster (Maybe Text)

The identifier (ID) of the cluster backup to restore. Use this value to restore the cluster from a backup instead of creating a new cluster. To find the backup ID, use DescribeBackups.

createCluster_hsmType :: Lens' CreateCluster Text

The type of HSM to use in the cluster. Currently the only allowed value is hsm1.medium.

createCluster_subnetIds :: Lens' CreateCluster (NonEmpty Text)

The identifiers (IDs) of the subnets where you are creating the cluster. You must specify at least one subnet. If you specify multiple subnets, they must meet the following criteria:

All subnets must be in the same virtual private cloud (VPC).

You can specify only one subnet per Availability Zone.

createClusterResponse_cluster :: Lens' CreateClusterResponse (Maybe Cluster)

Information about the cluster that was created.

createClusterResponse_httpStatus :: Lens' CreateClusterResponse Int

The response's http status code.

RestoreBackup

restoreBackup_backupId :: Lens' RestoreBackup Text

The ID of the backup to be restored. To find the ID of a backup, use the DescribeBackups operation.

restoreBackupResponse_backup :: Lens' RestoreBackupResponse (Maybe Backup)

Information on the Backup object created.

restoreBackupResponse_httpStatus :: Lens' RestoreBackupResponse Int

The response's http status code.

DeleteHsm

deleteHsm_eniId :: Lens' DeleteHsm (Maybe Text)

The identifier (ID) of the elastic network interface (ENI) of the HSM that you are deleting.

deleteHsm_hsmId :: Lens' DeleteHsm (Maybe Text)

The identifier (ID) of the HSM that you are deleting.

deleteHsm_eniIp :: Lens' DeleteHsm (Maybe Text)

The IP address of the elastic network interface (ENI) of the HSM that you are deleting.

deleteHsm_clusterId :: Lens' DeleteHsm Text

The identifier (ID) of the cluster that contains the HSM that you are deleting.

deleteHsmResponse_hsmId :: Lens' DeleteHsmResponse (Maybe Text)

The identifier (ID) of the HSM that was deleted.

deleteHsmResponse_httpStatus :: Lens' DeleteHsmResponse Int

The response's http status code.

ModifyCluster

modifyCluster_backupRetentionPolicy :: Lens' ModifyCluster BackupRetentionPolicy

A policy that defines how the service retains backups.

modifyCluster_clusterId :: Lens' ModifyCluster Text

The identifier (ID) of the cluster that you want to modify. To find the cluster ID, use DescribeClusters.

modifyClusterResponse_cluster :: Lens' ModifyClusterResponse (Maybe Cluster)

Undocumented member.

modifyClusterResponse_httpStatus :: Lens' ModifyClusterResponse Int

The response's http status code.

TagResource

tagResource_resourceId :: Lens' TagResource Text

The cluster identifier (ID) for the cluster that you are tagging. To find the cluster ID, use DescribeClusters.

tagResource_tagList :: Lens' TagResource [Tag]

A list of one or more tags.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

ListTags

listTags_nextToken :: Lens' ListTags (Maybe Text)

The NextToken value that you received in the previous response. Use this value to get more tags.

listTags_maxResults :: Lens' ListTags (Maybe Natural)

The maximum number of tags to return in the response. When there are more tags than the number you specify, the response contains a NextToken value.

listTags_resourceId :: Lens' ListTags Text

The cluster identifier (ID) for the cluster whose tags you are getting. To find the cluster ID, use DescribeClusters.

listTagsResponse_nextToken :: Lens' ListTagsResponse (Maybe Text)

An opaque string that indicates that the response contains only a subset of tags. Use this value in a subsequent ListTags request to get more tags.

listTagsResponse_httpStatus :: Lens' ListTagsResponse Int

The response's http status code.

listTagsResponse_tagList :: Lens' ListTagsResponse [Tag]

A list of tags.

UntagResource

untagResource_resourceId :: Lens' UntagResource Text

The cluster identifier (ID) for the cluster whose tags you are removing. To find the cluster ID, use DescribeClusters.

untagResource_tagKeyList :: Lens' UntagResource (NonEmpty Text)

A list of one or more tag keys for the tags that you are removing. Specify only the tag keys, not the tag values.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

ModifyBackupAttributes

modifyBackupAttributes_backupId :: Lens' ModifyBackupAttributes Text

The identifier (ID) of the backup to modify. To find the ID of a backup, use the DescribeBackups operation.

modifyBackupAttributes_neverExpires :: Lens' ModifyBackupAttributes Bool

Specifies whether the service should exempt a backup from the retention policy for the cluster. True exempts a backup from the retention policy. False means the service applies the backup retention policy defined at the cluster.

modifyBackupAttributesResponse_backup :: Lens' ModifyBackupAttributesResponse (Maybe Backup)

Undocumented member.

modifyBackupAttributesResponse_httpStatus :: Lens' ModifyBackupAttributesResponse Int

The response's http status code.

Types

Backup

backup_deleteTimestamp :: Lens' Backup (Maybe UTCTime)

The date and time when the backup will be permanently deleted.

backup_sourceCluster :: Lens' Backup (Maybe Text)

The identifier (ID) of the cluster containing the source backup from which the new backup was copied.

backup_neverExpires :: Lens' Backup (Maybe Bool)

Specifies whether the service should exempt a backup from the retention policy for the cluster. True exempts a backup from the retention policy. False means the service applies the backup retention policy defined at the cluster.

backup_sourceRegion :: Lens' Backup (Maybe Text)

The AWS Region that contains the source backup from which the new backup was copied.

backup_tagList :: Lens' Backup (Maybe [Tag])

The list of tags for the backup.

backup_sourceBackup :: Lens' Backup (Maybe Text)

The identifier (ID) of the source backup from which the new backup was copied.

backup_clusterId :: Lens' Backup (Maybe Text)

The identifier (ID) of the cluster that was backed up.

backup_createTimestamp :: Lens' Backup (Maybe UTCTime)

The date and time when the backup was created.

backup_copyTimestamp :: Lens' Backup (Maybe UTCTime)

The date and time when the backup was copied from a source backup.

backup_backupState :: Lens' Backup (Maybe BackupState)

The state of the backup.

backup_backupId :: Lens' Backup Text

The identifier (ID) of the backup.

BackupRetentionPolicy

backupRetentionPolicy_value :: Lens' BackupRetentionPolicy (Maybe Text)

Use a value between 7 - 379.

backupRetentionPolicy_type :: Lens' BackupRetentionPolicy (Maybe BackupRetentionType)

The type of backup retention policy. For the DAYS type, the value is the number of days to retain backups.

Certificates

certificates_manufacturerHardwareCertificate :: Lens' Certificates (Maybe Text)

The HSM hardware certificate issued (signed) by the hardware manufacturer.

certificates_clusterCsr :: Lens' Certificates (Maybe Text)

The cluster's certificate signing request (CSR). The CSR exists only when the cluster's state is UNINITIALIZED.

certificates_hsmCertificate :: Lens' Certificates (Maybe Text)

The HSM certificate issued (signed) by the HSM hardware.

certificates_clusterCertificate :: Lens' Certificates (Maybe Text)

The cluster certificate issued (signed) by the issuing certificate authority (CA) of the cluster's owner.

certificates_awsHardwareCertificate :: Lens' Certificates (Maybe Text)

The HSM hardware certificate issued (signed) by AWS CloudHSM.

Cluster

cluster_preCoPassword :: Lens' Cluster (Maybe Text)

The default password for the cluster's Pre-Crypto Officer (PRECO) user.

cluster_stateMessage :: Lens' Cluster (Maybe Text)

A description of the cluster's state.

cluster_state :: Lens' Cluster (Maybe ClusterState)

The cluster's state.

cluster_subnetMapping :: Lens' Cluster (Maybe (HashMap Text Text))

A map from availability zone to the cluster’s subnet in that availability zone.

cluster_backupRetentionPolicy :: Lens' Cluster (Maybe BackupRetentionPolicy)

A policy that defines how the service retains backups.

cluster_hsms :: Lens' Cluster (Maybe [Hsm])

Contains information about the HSMs in the cluster.

cluster_vpcId :: Lens' Cluster (Maybe Text)

The identifier (ID) of the virtual private cloud (VPC) that contains the cluster.

cluster_tagList :: Lens' Cluster (Maybe [Tag])

The list of tags for the cluster.

cluster_sourceBackupId :: Lens' Cluster (Maybe Text)

The identifier (ID) of the backup used to create the cluster. This value exists only when the cluster was created from a backup.

cluster_certificates :: Lens' Cluster (Maybe Certificates)

Contains one or more certificates or a certificate signing request (CSR).

cluster_securityGroup :: Lens' Cluster (Maybe Text)

The identifier (ID) of the cluster's security group.

cluster_clusterId :: Lens' Cluster (Maybe Text)

The cluster's identifier (ID).

cluster_createTimestamp :: Lens' Cluster (Maybe UTCTime)

The date and time when the cluster was created.

cluster_backupPolicy :: Lens' Cluster (Maybe BackupPolicy)

The cluster's backup policy.

cluster_hsmType :: Lens' Cluster (Maybe Text)

The type of HSM that the cluster contains.

DestinationBackup

destinationBackup_sourceCluster :: Lens' DestinationBackup (Maybe Text)

The identifier (ID) of the cluster containing the source backup from which the new backup was copied.

destinationBackup_sourceRegion :: Lens' DestinationBackup (Maybe Text)

The AWS region that contains the source backup from which the new backup was copied.

destinationBackup_sourceBackup :: Lens' DestinationBackup (Maybe Text)

The identifier (ID) of the source backup from which the new backup was copied.

destinationBackup_createTimestamp :: Lens' DestinationBackup (Maybe UTCTime)

The date and time when both the source backup was created.

Hsm

hsm_stateMessage :: Lens' Hsm (Maybe Text)

A description of the HSM's state.

hsm_state :: Lens' Hsm (Maybe HsmState)

The HSM's state.

hsm_eniId :: Lens' Hsm (Maybe Text)

The identifier (ID) of the HSM's elastic network interface (ENI).

hsm_subnetId :: Lens' Hsm (Maybe Text)

The subnet that contains the HSM's elastic network interface (ENI).

hsm_availabilityZone :: Lens' Hsm (Maybe Text)

The Availability Zone that contains the HSM.

hsm_clusterId :: Lens' Hsm (Maybe Text)

The identifier (ID) of the cluster that contains the HSM.

hsm_eniIp :: Lens' Hsm (Maybe Text)

The IP address of the HSM's elastic network interface (ENI).

hsm_hsmId :: Lens' Hsm Text

The HSM's identifier (ID).

Tag

tag_key :: Lens' Tag Text

The key of the tag.

tag_value :: Lens' Tag Text

The value of the tag.