Network.AWS.CloudHSMv2

Description

Derived from API version 2017-04-28 of the AWS service descriptions, licensed under Apache 2.0.

For more information about AWS CloudHSM, see AWS CloudHSM and the AWS CloudHSM User Guide.

Service Configuration

defaultService :: Service

API version 2017-04-28 of the Amazon CloudHSM V2 SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by CloudHSMv2.

CloudHsmInternalFailureException

_CloudHsmInternalFailureException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because of an AWS CloudHSM internal failure. The request can be retried.

CloudHsmResourceNotFoundException

_CloudHsmResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because it refers to a resource that cannot be found.

CloudHsmAccessDeniedException

_CloudHsmAccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because the requester does not have permission to perform the requested operation.

CloudHsmInvalidRequestException

_CloudHsmInvalidRequestException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because it is not a valid request.

CloudHsmServiceException

_CloudHsmServiceException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because an error occurred.

CloudHsmTagException

_CloudHsmTagException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because of a tagging failure. Verify the tag conditions in all applicable policies, and then retry the request.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

DeleteHsm

data DeleteHsm

See: newDeleteHsm smart constructor.

Constructors

DeleteHsm' (Maybe Text) (Maybe Text) (Maybe Text) Text

newDeleteHsm

Arguments

:: Text
$sel:clusterId:DeleteHsm'

-> DeleteHsm

Create a value of DeleteHsm with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eniIp:DeleteHsm', deleteHsm_eniIp - The IP address of the elastic network interface (ENI) of the HSM that you are deleting.

$sel:eniId:DeleteHsm', deleteHsm_eniId - The identifier (ID) of the elastic network interface (ENI) of the HSM that you are deleting.

$sel:hsmId:DeleteHsm', deleteHsm_hsmId - The identifier (ID) of the HSM that you are deleting.

$sel:clusterId:DeleteHsm', deleteHsm_clusterId - The identifier (ID) of the cluster that contains the HSM that you are deleting.

data DeleteHsmResponse

See: newDeleteHsmResponse smart constructor.

Constructors

DeleteHsmResponse' (Maybe Text) Int

newDeleteHsmResponse

Arguments

:: Int
$sel:httpStatus:DeleteHsmResponse'

-> DeleteHsmResponse

Create a value of DeleteHsmResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hsmId:DeleteHsm', deleteHsmResponse_hsmId - The identifier (ID) of the HSM that was deleted.

$sel:httpStatus:DeleteHsmResponse', deleteHsmResponse_httpStatus - The response's http status code.

DeleteBackup

data DeleteBackup

See: newDeleteBackup smart constructor.

Constructors

DeleteBackup' Text

newDeleteBackup

Arguments

:: Text
$sel:backupId:DeleteBackup'

-> DeleteBackup

Create a value of DeleteBackup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:backupId:DeleteBackup', deleteBackup_backupId - The ID of the backup to be deleted. To find the ID of a backup, use the DescribeBackups operation.

data DeleteBackupResponse

See: newDeleteBackupResponse smart constructor.

Constructors

DeleteBackupResponse' (Maybe Backup) Int

newDeleteBackupResponse

Arguments

:: Int
$sel:httpStatus:DeleteBackupResponse'

-> DeleteBackupResponse

Create a value of DeleteBackupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:backup:DeleteBackupResponse', deleteBackupResponse_backup - Information on the Backup object deleted.

$sel:httpStatus:DeleteBackupResponse', deleteBackupResponse_httpStatus - The response's http status code.

DescribeClusters (Paginated)

data DescribeClusters

See: newDescribeClusters smart constructor.

Constructors

DescribeClusters' (Maybe Text) (Maybe Natural) (Maybe (HashMap Text [Text]))

newDescribeClusters :: DescribeClusters

Create a value of DescribeClusters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeClusters', describeClusters_nextToken - The NextToken value that you received in the previous response. Use this value to get more clusters.

$sel:maxResults:DescribeClusters', describeClusters_maxResults - The maximum number of clusters to return in the response. When there are more clusters than the number you specify, the response contains a NextToken value.

$sel:filters:DescribeClusters', describeClusters_filters - One or more filters to limit the items returned in the response.

Use the clusterIds filter to return only the specified clusters. Specify clusters by their cluster identifier (ID).

Use the vpcIds filter to return only the clusters in the specified virtual private clouds (VPCs). Specify VPCs by their VPC identifier (ID).

Use the states filter to return only clusters that match the specified state.

data DescribeClustersResponse

See: newDescribeClustersResponse smart constructor.

Constructors

DescribeClustersResponse' (Maybe Text) (Maybe [Cluster]) Int

newDescribeClustersResponse

Arguments

:: Int
$sel:httpStatus:DescribeClustersResponse'

-> DescribeClustersResponse

Create a value of DescribeClustersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeClusters', describeClustersResponse_nextToken - An opaque string that indicates that the response contains only a subset of clusters. Use this value in a subsequent DescribeClusters request to get more clusters.

$sel:clusters:DescribeClustersResponse', describeClustersResponse_clusters - A list of clusters.

$sel:httpStatus:DescribeClustersResponse', describeClustersResponse_httpStatus - The response's http status code.

RestoreBackup

data RestoreBackup

See: newRestoreBackup smart constructor.

Constructors

RestoreBackup' Text

newRestoreBackup

Arguments

:: Text
$sel:backupId:RestoreBackup'

-> RestoreBackup

Create a value of RestoreBackup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:backupId:RestoreBackup', restoreBackup_backupId - The ID of the backup to be restored. To find the ID of a backup, use the DescribeBackups operation.

data RestoreBackupResponse

See: newRestoreBackupResponse smart constructor.

Constructors

RestoreBackupResponse' (Maybe Backup) Int

newRestoreBackupResponse

Arguments

:: Int
$sel:httpStatus:RestoreBackupResponse'

-> RestoreBackupResponse

Create a value of RestoreBackupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:backup:RestoreBackupResponse', restoreBackupResponse_backup - Information on the Backup object created.

$sel:httpStatus:RestoreBackupResponse', restoreBackupResponse_httpStatus - The response's http status code.

CreateCluster

data CreateCluster

See: newCreateCluster smart constructor.

Constructors

CreateCluster' (Maybe Text) (Maybe [Tag]) (Maybe BackupRetentionPolicy) Text (NonEmpty Text)

newCreateCluster

Arguments

:: Text
$sel:hsmType:CreateCluster'

-> NonEmpty Text
$sel:subnetIds:CreateCluster'

-> CreateCluster

Create a value of CreateCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceBackupId:CreateCluster', createCluster_sourceBackupId - The identifier (ID) of the cluster backup to restore. Use this value to restore the cluster from a backup instead of creating a new cluster. To find the backup ID, use DescribeBackups.

$sel:tagList:CreateCluster', createCluster_tagList - Tags to apply to the CloudHSM cluster during creation.

$sel:backupRetentionPolicy:CreateCluster', createCluster_backupRetentionPolicy - A policy that defines how the service retains backups.

$sel:hsmType:CreateCluster', createCluster_hsmType - The type of HSM to use in the cluster. Currently the only allowed value is hsm1.medium.

$sel:subnetIds:CreateCluster', createCluster_subnetIds - The identifiers (IDs) of the subnets where you are creating the cluster. You must specify at least one subnet. If you specify multiple subnets, they must meet the following criteria:

All subnets must be in the same virtual private cloud (VPC).

You can specify only one subnet per Availability Zone.

data CreateClusterResponse

See: newCreateClusterResponse smart constructor.

Constructors

CreateClusterResponse' (Maybe Cluster) Int

newCreateClusterResponse

Arguments

:: Int
$sel:httpStatus:CreateClusterResponse'

-> CreateClusterResponse

Create a value of CreateClusterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cluster:CreateClusterResponse', createClusterResponse_cluster - Information about the cluster that was created.

$sel:httpStatus:CreateClusterResponse', createClusterResponse_httpStatus - The response's http status code.

UntagResource

data UntagResource

See: newUntagResource smart constructor.

Constructors

UntagResource' Text (NonEmpty Text)

newUntagResource

Arguments

:: Text
$sel:resourceId:UntagResource'

-> NonEmpty Text
$sel:tagKeyList:UntagResource'

-> UntagResource

Create a value of UntagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceId:UntagResource', untagResource_resourceId - The cluster identifier (ID) for the cluster whose tags you are removing. To find the cluster ID, use DescribeClusters.

$sel:tagKeyList:UntagResource', untagResource_tagKeyList - A list of one or more tag keys for the tags that you are removing. Specify only the tag keys, not the tag values.

data UntagResourceResponse

See: newUntagResourceResponse smart constructor.

Constructors

UntagResourceResponse' Int

newUntagResourceResponse

Arguments

:: Int
$sel:httpStatus:UntagResourceResponse'

-> UntagResourceResponse

Create a value of UntagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UntagResourceResponse', untagResourceResponse_httpStatus - The response's http status code.

TagResource

data TagResource

See: newTagResource smart constructor.

Constructors

TagResource' Text [Tag]

newTagResource

Arguments

:: Text
$sel:resourceId:TagResource'

-> TagResource

Create a value of TagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceId:TagResource', tagResource_resourceId - The cluster identifier (ID) for the cluster that you are tagging. To find the cluster ID, use DescribeClusters.

$sel:tagList:TagResource', tagResource_tagList - A list of one or more tags.

data TagResourceResponse

See: newTagResourceResponse smart constructor.

Constructors

TagResourceResponse' Int

newTagResourceResponse

Arguments

:: Int
$sel:httpStatus:TagResourceResponse'

-> TagResourceResponse

Create a value of TagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:TagResourceResponse', tagResourceResponse_httpStatus - The response's http status code.

CopyBackupToRegion

data CopyBackupToRegion

See: newCopyBackupToRegion smart constructor.

Constructors

CopyBackupToRegion' (Maybe [Tag]) Text Text

newCopyBackupToRegion

Arguments

:: Text
$sel:destinationRegion:CopyBackupToRegion'

-> Text
$sel:backupId:CopyBackupToRegion'

-> CopyBackupToRegion

Create a value of CopyBackupToRegion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tagList:CopyBackupToRegion', copyBackupToRegion_tagList - Tags to apply to the destination backup during creation. If you specify tags, only these tags will be applied to the destination backup. If you do not specify tags, the service copies tags from the source backup to the destination backup.

$sel:destinationRegion:CopyBackupToRegion', copyBackupToRegion_destinationRegion - The AWS region that will contain your copied CloudHSM cluster backup.

$sel:backupId:CopyBackupToRegion', copyBackupToRegion_backupId - The ID of the backup that will be copied to the destination region.

data CopyBackupToRegionResponse

See: newCopyBackupToRegionResponse smart constructor.

Constructors

CopyBackupToRegionResponse' (Maybe DestinationBackup) Int

newCopyBackupToRegionResponse

Arguments

:: Int
$sel:httpStatus:CopyBackupToRegionResponse'

-> CopyBackupToRegionResponse

Create a value of CopyBackupToRegionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:destinationBackup:CopyBackupToRegionResponse', copyBackupToRegionResponse_destinationBackup - Information on the backup that will be copied to the destination region, including CreateTimestamp, SourceBackup, SourceCluster, and Source Region. CreateTimestamp of the destination backup will be the same as that of the source backup.

You will need to use the sourceBackupID returned in this operation to use the DescribeBackups operation on the backup that will be copied to the destination region.

$sel:httpStatus:CopyBackupToRegionResponse', copyBackupToRegionResponse_httpStatus - The response's http status code.

ModifyCluster

data ModifyCluster

See: newModifyCluster smart constructor.

Constructors

ModifyCluster' BackupRetentionPolicy Text

newModifyCluster

Arguments

:: BackupRetentionPolicy
$sel:backupRetentionPolicy:ModifyCluster'

-> Text
$sel:clusterId:ModifyCluster'

-> ModifyCluster

Create a value of ModifyCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:backupRetentionPolicy:ModifyCluster', modifyCluster_backupRetentionPolicy - A policy that defines how the service retains backups.

$sel:clusterId:ModifyCluster', modifyCluster_clusterId - The identifier (ID) of the cluster that you want to modify. To find the cluster ID, use DescribeClusters.

data ModifyClusterResponse

See: newModifyClusterResponse smart constructor.

Constructors

ModifyClusterResponse' (Maybe Cluster) Int

newModifyClusterResponse

Arguments

:: Int
$sel:httpStatus:ModifyClusterResponse'

-> ModifyClusterResponse

Create a value of ModifyClusterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cluster:ModifyClusterResponse', modifyClusterResponse_cluster - Undocumented member.

$sel:httpStatus:ModifyClusterResponse', modifyClusterResponse_httpStatus - The response's http status code.

ModifyBackupAttributes

data ModifyBackupAttributes

See: newModifyBackupAttributes smart constructor.

Constructors

ModifyBackupAttributes' Text Bool

newModifyBackupAttributes

Arguments

:: Text
$sel:backupId:ModifyBackupAttributes'

-> Bool
$sel:neverExpires:ModifyBackupAttributes'

-> ModifyBackupAttributes

Create a value of ModifyBackupAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:backupId:ModifyBackupAttributes', modifyBackupAttributes_backupId - The identifier (ID) of the backup to modify. To find the ID of a backup, use the DescribeBackups operation.

$sel:neverExpires:ModifyBackupAttributes', modifyBackupAttributes_neverExpires - Specifies whether the service should exempt a backup from the retention policy for the cluster. True exempts a backup from the retention policy. False means the service applies the backup retention policy defined at the cluster.

data ModifyBackupAttributesResponse

See: newModifyBackupAttributesResponse smart constructor.

Constructors

ModifyBackupAttributesResponse' (Maybe Backup) Int

newModifyBackupAttributesResponse

Arguments

:: Int
$sel:httpStatus:ModifyBackupAttributesResponse'

-> ModifyBackupAttributesResponse

Create a value of ModifyBackupAttributesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:backup:ModifyBackupAttributesResponse', modifyBackupAttributesResponse_backup - Undocumented member.

$sel:httpStatus:ModifyBackupAttributesResponse', modifyBackupAttributesResponse_httpStatus - The response's http status code.

ListTags (Paginated)

data ListTags

See: newListTags smart constructor.

Constructors

ListTags' (Maybe Text) (Maybe Natural) Text

newListTags

Arguments

:: Text
$sel:resourceId:ListTags'

-> ListTags

Create a value of ListTags with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListTags', listTags_nextToken - The NextToken value that you received in the previous response. Use this value to get more tags.

$sel:maxResults:ListTags', listTags_maxResults - The maximum number of tags to return in the response. When there are more tags than the number you specify, the response contains a NextToken value.

$sel:resourceId:ListTags', listTags_resourceId - The cluster identifier (ID) for the cluster whose tags you are getting. To find the cluster ID, use DescribeClusters.

data ListTagsResponse

See: newListTagsResponse smart constructor.

Constructors

ListTagsResponse' (Maybe Text) Int [Tag]

newListTagsResponse

Arguments

:: Int
$sel:httpStatus:ListTagsResponse'

-> ListTagsResponse

Create a value of ListTagsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListTags', listTagsResponse_nextToken - An opaque string that indicates that the response contains only a subset of tags. Use this value in a subsequent ListTags request to get more tags.

$sel:httpStatus:ListTagsResponse', listTagsResponse_httpStatus - The response's http status code.

$sel:tagList:ListTagsResponse', listTagsResponse_tagList - A list of tags.

DeleteCluster

data DeleteCluster

See: newDeleteCluster smart constructor.

Constructors

DeleteCluster' Text

newDeleteCluster

Arguments

:: Text
$sel:clusterId:DeleteCluster'

-> DeleteCluster

Create a value of DeleteCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clusterId:DeleteCluster', deleteCluster_clusterId - The identifier (ID) of the cluster that you are deleting. To find the cluster ID, use DescribeClusters.

data DeleteClusterResponse

See: newDeleteClusterResponse smart constructor.

Constructors

DeleteClusterResponse' (Maybe Cluster) Int

newDeleteClusterResponse

Arguments

:: Int
$sel:httpStatus:DeleteClusterResponse'

-> DeleteClusterResponse

Create a value of DeleteClusterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cluster:DeleteClusterResponse', deleteClusterResponse_cluster - Information about the cluster that was deleted.

$sel:httpStatus:DeleteClusterResponse', deleteClusterResponse_httpStatus - The response's http status code.

DescribeBackups (Paginated)

data DescribeBackups

See: newDescribeBackups smart constructor.

Constructors

DescribeBackups' (Maybe Text) (Maybe Natural) (Maybe Bool) (Maybe (HashMap Text [Text]))

newDescribeBackups :: DescribeBackups

Create a value of DescribeBackups with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeBackups', describeBackups_nextToken - The NextToken value that you received in the previous response. Use this value to get more backups.

$sel:maxResults:DescribeBackups', describeBackups_maxResults - The maximum number of backups to return in the response. When there are more backups than the number you specify, the response contains a NextToken value.

$sel:sortAscending:DescribeBackups', describeBackups_sortAscending - Designates whether or not to sort the return backups by ascending chronological order of generation.

$sel:filters:DescribeBackups', describeBackups_filters - One or more filters to limit the items returned in the response.

Use the backupIds filter to return only the specified backups. Specify backups by their backup identifier (ID).

Use the sourceBackupIds filter to return only the backups created from a source backup. The sourceBackupID of a source backup is returned by the CopyBackupToRegion operation.

Use the clusterIds filter to return only the backups for the specified clusters. Specify clusters by their cluster identifier (ID).

Use the states filter to return only backups that match the specified state.

Use the neverExpires filter to return backups filtered by the value in the neverExpires parameter. True returns all backups exempt from the backup retention policy. False returns all backups with a backup retention policy defined at the cluster.

data DescribeBackupsResponse

See: newDescribeBackupsResponse smart constructor.

Constructors

DescribeBackupsResponse' (Maybe Text) (Maybe [Backup]) Int

newDescribeBackupsResponse

Arguments

:: Int
$sel:httpStatus:DescribeBackupsResponse'

-> DescribeBackupsResponse

Create a value of DescribeBackupsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeBackups', describeBackupsResponse_nextToken - An opaque string that indicates that the response contains only a subset of backups. Use this value in a subsequent DescribeBackups request to get more backups.

$sel:backups:DescribeBackupsResponse', describeBackupsResponse_backups - A list of backups.

$sel:httpStatus:DescribeBackupsResponse', describeBackupsResponse_httpStatus - The response's http status code.

InitializeCluster

data InitializeCluster

See: newInitializeCluster smart constructor.

Constructors

InitializeCluster' Text Text Text

newInitializeCluster

Arguments

:: Text
$sel:clusterId:InitializeCluster'

-> Text
$sel:signedCert:InitializeCluster'

-> Text
$sel:trustAnchor:InitializeCluster'

-> InitializeCluster

Create a value of InitializeCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clusterId:InitializeCluster', initializeCluster_clusterId - The identifier (ID) of the cluster that you are claiming. To find the cluster ID, use DescribeClusters.

$sel:signedCert:InitializeCluster', initializeCluster_signedCert - The cluster certificate issued (signed) by your issuing certificate authority (CA). The certificate must be in PEM format and can contain a maximum of 5000 characters.

$sel:trustAnchor:InitializeCluster', initializeCluster_trustAnchor - The issuing certificate of the issuing certificate authority (CA) that issued (signed) the cluster certificate. You must use a self-signed certificate. The certificate used to sign the HSM CSR must be directly available, and thus must be the root certificate. The certificate must be in PEM format and can contain a maximum of 5000 characters.

data InitializeClusterResponse

See: newInitializeClusterResponse smart constructor.

Constructors

InitializeClusterResponse' (Maybe Text) (Maybe ClusterState) Int

newInitializeClusterResponse

Arguments

:: Int
$sel:httpStatus:InitializeClusterResponse'

-> InitializeClusterResponse

Create a value of InitializeClusterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stateMessage:InitializeClusterResponse', initializeClusterResponse_stateMessage - A description of the cluster's state.

$sel:state:InitializeClusterResponse', initializeClusterResponse_state - The cluster's state.

$sel:httpStatus:InitializeClusterResponse', initializeClusterResponse_httpStatus - The response's http status code.

CreateHsm

data CreateHsm

See: newCreateHsm smart constructor.

Constructors

CreateHsm' (Maybe Text) Text Text

newCreateHsm

Arguments

:: Text
$sel:clusterId:CreateHsm'

-> Text
$sel:availabilityZone:CreateHsm'

-> CreateHsm

Create a value of CreateHsm with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ipAddress:CreateHsm', createHsm_ipAddress - The HSM's IP address. If you specify an IP address, use an available address from the subnet that maps to the Availability Zone where you are creating the HSM. If you don't specify an IP address, one is chosen for you from that subnet.

$sel:clusterId:CreateHsm', createHsm_clusterId - The identifier (ID) of the HSM's cluster. To find the cluster ID, use DescribeClusters.

$sel:availabilityZone:CreateHsm', createHsm_availabilityZone - The Availability Zone where you are creating the HSM. To find the cluster's Availability Zones, use DescribeClusters.

data CreateHsmResponse

See: newCreateHsmResponse smart constructor.

Constructors

CreateHsmResponse' (Maybe Hsm) Int

newCreateHsmResponse

Arguments

:: Int
$sel:httpStatus:CreateHsmResponse'

-> CreateHsmResponse

Create a value of CreateHsmResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hsm:CreateHsmResponse', createHsmResponse_hsm - Information about the HSM that was created.

$sel:httpStatus:CreateHsmResponse', createHsmResponse_httpStatus - The response's http status code.

Types

BackupPolicy

newtype BackupPolicy

Constructors

BackupPolicy'

Fields

fromBackupPolicy :: Text

Bundled Patterns

pattern BackupPolicy_DEFAULT :: BackupPolicy

BackupRetentionType

newtype BackupRetentionType

Constructors

BackupRetentionType'

Fields

fromBackupRetentionType :: Text

Bundled Patterns

pattern BackupRetentionType_DAYS :: BackupRetentionType

BackupState

newtype BackupState

Constructors

BackupState'

Fields

fromBackupState :: Text

Bundled Patterns

pattern BackupState_CREATE_IN_PROGRESS :: BackupState

pattern BackupState_DELETED :: BackupState

pattern BackupState_PENDING_DELETION :: BackupState

pattern BackupState_READY :: BackupState

ClusterState

newtype ClusterState

Constructors

ClusterState'

Fields

fromClusterState :: Text

Bundled Patterns

pattern ClusterState_ACTIVE :: ClusterState

pattern ClusterState_CREATE_IN_PROGRESS :: ClusterState

pattern ClusterState_DEGRADED :: ClusterState

pattern ClusterState_DELETED :: ClusterState

pattern ClusterState_DELETE_IN_PROGRESS :: ClusterState

pattern ClusterState_INITIALIZED :: ClusterState

pattern ClusterState_INITIALIZE_IN_PROGRESS :: ClusterState

pattern ClusterState_UNINITIALIZED :: ClusterState

pattern ClusterState_UPDATE_IN_PROGRESS :: ClusterState

HsmState

newtype HsmState

Constructors

HsmState'

Fields

fromHsmState :: Text

Bundled Patterns

pattern HsmState_ACTIVE :: HsmState

pattern HsmState_CREATE_IN_PROGRESS :: HsmState

pattern HsmState_DEGRADED :: HsmState

pattern HsmState_DELETED :: HsmState

pattern HsmState_DELETE_IN_PROGRESS :: HsmState

Backup

data Backup

Contains information about a backup of an AWS CloudHSM cluster. All backup objects contain the BackupId, BackupState, ClusterId, and CreateTimestamp parameters. Backups that were copied into a destination region additionally contain the CopyTimestamp, SourceBackup, SourceCluster, and SourceRegion parameters. A backup that is pending deletion will include the DeleteTimestamp parameter.

See: newBackup smart constructor.

Constructors

Backup' (Maybe Text) (Maybe BackupState) (Maybe Text) (Maybe POSIX) (Maybe Bool) (Maybe POSIX) (Maybe POSIX) (Maybe Text) (Maybe [Tag]) (Maybe Text) Text

newBackup

Arguments

:: Text
$sel:backupId:Backup'

-> Backup

Create a value of Backup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clusterId:Backup', backup_clusterId - The identifier (ID) of the cluster that was backed up.

$sel:backupState:Backup', backup_backupState - The state of the backup.

$sel:sourceBackup:Backup', backup_sourceBackup - The identifier (ID) of the source backup from which the new backup was copied.

$sel:copyTimestamp:Backup', backup_copyTimestamp - The date and time when the backup was copied from a source backup.

$sel:neverExpires:Backup', backup_neverExpires - Specifies whether the service should exempt a backup from the retention policy for the cluster. True exempts a backup from the retention policy. False means the service applies the backup retention policy defined at the cluster.

$sel:createTimestamp:Backup', backup_createTimestamp - The date and time when the backup was created.

$sel:deleteTimestamp:Backup', backup_deleteTimestamp - The date and time when the backup will be permanently deleted.

$sel:sourceCluster:Backup', backup_sourceCluster - The identifier (ID) of the cluster containing the source backup from which the new backup was copied.

$sel:tagList:Backup', backup_tagList - The list of tags for the backup.

$sel:sourceRegion:Backup', backup_sourceRegion - The AWS Region that contains the source backup from which the new backup was copied.

$sel:backupId:Backup', backup_backupId - The identifier (ID) of the backup.

BackupRetentionPolicy

data BackupRetentionPolicy

A policy that defines the number of days to retain backups.

See: newBackupRetentionPolicy smart constructor.

Constructors

BackupRetentionPolicy' (Maybe Text) (Maybe BackupRetentionType)

newBackupRetentionPolicy :: BackupRetentionPolicy

Create a value of BackupRetentionPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:BackupRetentionPolicy', backupRetentionPolicy_value - Use a value between 7 - 379.

$sel:type':BackupRetentionPolicy', backupRetentionPolicy_type - The type of backup retention policy. For the DAYS type, the value is the number of days to retain backups.

Certificates

data Certificates

Contains one or more certificates or a certificate signing request (CSR).

See: newCertificates smart constructor.

Constructors

Certificates' (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text)

newCertificates :: Certificates

Create a value of Certificates with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:awsHardwareCertificate:Certificates', certificates_awsHardwareCertificate - The HSM hardware certificate issued (signed) by AWS CloudHSM.

$sel:hsmCertificate:Certificates', certificates_hsmCertificate - The HSM certificate issued (signed) by the HSM hardware.

$sel:clusterCsr:Certificates', certificates_clusterCsr - The cluster's certificate signing request (CSR). The CSR exists only when the cluster's state is UNINITIALIZED.

$sel:clusterCertificate:Certificates', certificates_clusterCertificate - The cluster certificate issued (signed) by the issuing certificate authority (CA) of the cluster's owner.

$sel:manufacturerHardwareCertificate:Certificates', certificates_manufacturerHardwareCertificate - The HSM hardware certificate issued (signed) by the hardware manufacturer.

Cluster

data Cluster

Contains information about an AWS CloudHSM cluster.

See: newCluster smart constructor.

Constructors

Cluster' (Maybe Text) (Maybe Text) (Maybe BackupPolicy) (Maybe POSIX) (Maybe (HashMap Text Text)) (Maybe ClusterState) (Maybe Text) (Maybe Text) (Maybe Certificates) (Maybe Text) (Maybe Text) (Maybe [Tag]) (Maybe Text) (Maybe [Hsm]) (Maybe BackupRetentionPolicy)

newCluster :: Cluster

Create a value of Cluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clusterId:Cluster', cluster_clusterId - The cluster's identifier (ID).

$sel:stateMessage:Cluster', cluster_stateMessage - A description of the cluster's state.

$sel:backupPolicy:Cluster', cluster_backupPolicy - The cluster's backup policy.

$sel:createTimestamp:Cluster', cluster_createTimestamp - The date and time when the cluster was created.

$sel:subnetMapping:Cluster', cluster_subnetMapping - A map from availability zone to the cluster’s subnet in that availability zone.

$sel:state:Cluster', cluster_state - The cluster's state.

$sel:preCoPassword:Cluster', cluster_preCoPassword - The default password for the cluster's Pre-Crypto Officer (PRECO) user.

$sel:securityGroup:Cluster', cluster_securityGroup - The identifier (ID) of the cluster's security group.

$sel:certificates:Cluster', cluster_certificates - Contains one or more certificates or a certificate signing request (CSR).

$sel:sourceBackupId:Cluster', cluster_sourceBackupId - The identifier (ID) of the backup used to create the cluster. This value exists only when the cluster was created from a backup.

$sel:hsmType:Cluster', cluster_hsmType - The type of HSM that the cluster contains.

$sel:tagList:Cluster', cluster_tagList - The list of tags for the cluster.

$sel:vpcId:Cluster', cluster_vpcId - The identifier (ID) of the virtual private cloud (VPC) that contains the cluster.

$sel:hsms:Cluster', cluster_hsms - Contains information about the HSMs in the cluster.

$sel:backupRetentionPolicy:Cluster', cluster_backupRetentionPolicy - A policy that defines how the service retains backups.

DestinationBackup

data DestinationBackup

Contains information about the backup that will be copied and created by the CopyBackupToRegion operation.

See: newDestinationBackup smart constructor.

Constructors

DestinationBackup' (Maybe Text) (Maybe POSIX) (Maybe Text) (Maybe Text)

newDestinationBackup :: DestinationBackup

Create a value of DestinationBackup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceBackup:DestinationBackup', destinationBackup_sourceBackup - The identifier (ID) of the source backup from which the new backup was copied.

$sel:createTimestamp:DestinationBackup', destinationBackup_createTimestamp - The date and time when both the source backup was created.

$sel:sourceCluster:DestinationBackup', destinationBackup_sourceCluster - The identifier (ID) of the cluster containing the source backup from which the new backup was copied.

$sel:sourceRegion:DestinationBackup', destinationBackup_sourceRegion - The AWS region that contains the source backup from which the new backup was copied.

Hsm

data Hsm

Contains information about a hardware security module (HSM) in an AWS CloudHSM cluster.

See: newHsm smart constructor.

Constructors

Hsm' (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe HsmState) (Maybe Text) Text

newHsm

Arguments

:: Text
$sel:hsmId:Hsm'

-> Hsm

Create a value of Hsm with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clusterId:Hsm', hsm_clusterId - The identifier (ID) of the cluster that contains the HSM.

$sel:stateMessage:Hsm', hsm_stateMessage - A description of the HSM's state.

$sel:eniIp:Hsm', hsm_eniIp - The IP address of the HSM's elastic network interface (ENI).

$sel:eniId:Hsm', hsm_eniId - The identifier (ID) of the HSM's elastic network interface (ENI).

$sel:availabilityZone:Hsm', hsm_availabilityZone - The Availability Zone that contains the HSM.

$sel:state:Hsm', hsm_state - The HSM's state.

$sel:subnetId:Hsm', hsm_subnetId - The subnet that contains the HSM's elastic network interface (ENI).

$sel:hsmId:Hsm', hsm_hsmId - The HSM's identifier (ID).

Tag

data Tag

Contains a tag. A tag is a key-value pair.

See: newTag smart constructor.

Constructors

Tag' Text Text

newTag

Arguments

:: Text
$sel:key:Tag'

-> Text
$sel:value:Tag'

-> Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:Tag', tag_key - The key of the tag.

$sel:value:Tag', tag_value - The value of the tag.