Network.AWS.CloudHSMv2.CreateHsm

Description

Creates a new hardware security module (HSM) in the specified AWS CloudHSM cluster.

Creating a Request

data CreateHsm

See: newCreateHsm smart constructor.

Constructors

CreateHsm'

Fields

ipAddress :: Maybe Text

The HSM's IP address. If you specify an IP address, use an available address from the subnet that maps to the Availability Zone where you are creating the HSM. If you don't specify an IP address, one is chosen for you from that subnet.

clusterId :: Text

The identifier (ID) of the HSM's cluster. To find the cluster ID, use DescribeClusters.

availabilityZone :: Text

The Availability Zone where you are creating the HSM. To find the cluster's Availability Zones, use DescribeClusters.

newCreateHsm

Arguments

:: Text
$sel:clusterId:CreateHsm'

-> Text
$sel:availabilityZone:CreateHsm'

-> CreateHsm

Create a value of CreateHsm with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ipAddress:CreateHsm', createHsm_ipAddress - The HSM's IP address. If you specify an IP address, use an available address from the subnet that maps to the Availability Zone where you are creating the HSM. If you don't specify an IP address, one is chosen for you from that subnet.

$sel:clusterId:CreateHsm', createHsm_clusterId - The identifier (ID) of the HSM's cluster. To find the cluster ID, use DescribeClusters.

$sel:availabilityZone:CreateHsm', createHsm_availabilityZone - The Availability Zone where you are creating the HSM. To find the cluster's Availability Zones, use DescribeClusters.

Request Lenses

createHsm_ipAddress :: Lens' CreateHsm (Maybe Text)

The HSM's IP address. If you specify an IP address, use an available address from the subnet that maps to the Availability Zone where you are creating the HSM. If you don't specify an IP address, one is chosen for you from that subnet.

createHsm_clusterId :: Lens' CreateHsm Text

The identifier (ID) of the HSM's cluster. To find the cluster ID, use DescribeClusters.

createHsm_availabilityZone :: Lens' CreateHsm Text

The Availability Zone where you are creating the HSM. To find the cluster's Availability Zones, use DescribeClusters.

Destructuring the Response

data CreateHsmResponse

See: newCreateHsmResponse smart constructor.

Constructors

CreateHsmResponse'

Fields

hsm :: Maybe Hsm

Information about the HSM that was created.

httpStatus :: Int

The response's http status code.

newCreateHsmResponse

Arguments

:: Int
$sel:httpStatus:CreateHsmResponse'

-> CreateHsmResponse

Create a value of CreateHsmResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hsm:CreateHsmResponse', createHsmResponse_hsm - Information about the HSM that was created.

$sel:httpStatus:CreateHsmResponse', createHsmResponse_httpStatus - The response's http status code.

Response Lenses

createHsmResponse_hsm :: Lens' CreateHsmResponse (Maybe Hsm)

Information about the HSM that was created.

createHsmResponse_httpStatus :: Lens' CreateHsmResponse Int

The response's http status code.