Network.AWS.CloudHSMv2.DescribeBackups

Description

Gets information about backups of AWS CloudHSM clusters.

This is a paginated operation, which means that each response might contain only a subset of all the backups. When the response contains only a subset of backups, it includes a NextToken value. Use this value in a subsequent DescribeBackups request to get more backups. When you receive a response with no NextToken (or an empty or null value), that means there are no more backups to get.

This operation returns paginated results.

Creating a Request

data DescribeBackups

See: newDescribeBackups smart constructor.

Constructors

DescribeBackups'

Fields

sortAscending :: Maybe Bool

Designates whether or not to sort the return backups by ascending chronological order of generation.

filters :: Maybe (HashMap Text [Text])

One or more filters to limit the items returned in the response.

Use the backupIds filter to return only the specified backups. Specify backups by their backup identifier (ID).

Use the sourceBackupIds filter to return only the backups created from a source backup. The sourceBackupID of a source backup is returned by the CopyBackupToRegion operation.

Use the clusterIds filter to return only the backups for the specified clusters. Specify clusters by their cluster identifier (ID).

Use the states filter to return only backups that match the specified state.

Use the neverExpires filter to return backups filtered by the value in the neverExpires parameter. True returns all backups exempt from the backup retention policy. False returns all backups with a backup retention policy defined at the cluster.

nextToken :: Maybe Text

The NextToken value that you received in the previous response. Use this value to get more backups.

maxResults :: Maybe Natural

The maximum number of backups to return in the response. When there are more backups than the number you specify, the response contains a NextToken value.

newDescribeBackups :: DescribeBackups

Create a value of DescribeBackups with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sortAscending:DescribeBackups', describeBackups_sortAscending - Designates whether or not to sort the return backups by ascending chronological order of generation.

$sel:filters:DescribeBackups', describeBackups_filters - One or more filters to limit the items returned in the response.

Use the backupIds filter to return only the specified backups. Specify backups by their backup identifier (ID).

Use the sourceBackupIds filter to return only the backups created from a source backup. The sourceBackupID of a source backup is returned by the CopyBackupToRegion operation.

Use the clusterIds filter to return only the backups for the specified clusters. Specify clusters by their cluster identifier (ID).

Use the states filter to return only backups that match the specified state.

Use the neverExpires filter to return backups filtered by the value in the neverExpires parameter. True returns all backups exempt from the backup retention policy. False returns all backups with a backup retention policy defined at the cluster.

$sel:nextToken:DescribeBackups', describeBackups_nextToken - The NextToken value that you received in the previous response. Use this value to get more backups.

$sel:maxResults:DescribeBackups', describeBackups_maxResults - The maximum number of backups to return in the response. When there are more backups than the number you specify, the response contains a NextToken value.

Request Lenses

describeBackups_sortAscending :: Lens' DescribeBackups (Maybe Bool)

Designates whether or not to sort the return backups by ascending chronological order of generation.

describeBackups_filters :: Lens' DescribeBackups (Maybe (HashMap Text [Text]))

One or more filters to limit the items returned in the response.

Use the backupIds filter to return only the specified backups. Specify backups by their backup identifier (ID).

Use the sourceBackupIds filter to return only the backups created from a source backup. The sourceBackupID of a source backup is returned by the CopyBackupToRegion operation.

Use the clusterIds filter to return only the backups for the specified clusters. Specify clusters by their cluster identifier (ID).

Use the states filter to return only backups that match the specified state.

Use the neverExpires filter to return backups filtered by the value in the neverExpires parameter. True returns all backups exempt from the backup retention policy. False returns all backups with a backup retention policy defined at the cluster.

describeBackups_nextToken :: Lens' DescribeBackups (Maybe Text)

The NextToken value that you received in the previous response. Use this value to get more backups.

describeBackups_maxResults :: Lens' DescribeBackups (Maybe Natural)

The maximum number of backups to return in the response. When there are more backups than the number you specify, the response contains a NextToken value.

Destructuring the Response

data DescribeBackupsResponse

See: newDescribeBackupsResponse smart constructor.

Constructors

DescribeBackupsResponse'

Fields

backups :: Maybe [Backup]

A list of backups.

nextToken :: Maybe Text

An opaque string that indicates that the response contains only a subset of backups. Use this value in a subsequent DescribeBackups request to get more backups.

httpStatus :: Int

The response's http status code.

newDescribeBackupsResponse

Arguments

:: Int
$sel:httpStatus:DescribeBackupsResponse'

-> DescribeBackupsResponse

Create a value of DescribeBackupsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:backups:DescribeBackupsResponse', describeBackupsResponse_backups - A list of backups.

$sel:nextToken:DescribeBackups', describeBackupsResponse_nextToken - An opaque string that indicates that the response contains only a subset of backups. Use this value in a subsequent DescribeBackups request to get more backups.

$sel:httpStatus:DescribeBackupsResponse', describeBackupsResponse_httpStatus - The response's http status code.

Response Lenses

describeBackupsResponse_backups :: Lens' DescribeBackupsResponse (Maybe [Backup])

A list of backups.

describeBackupsResponse_nextToken :: Lens' DescribeBackupsResponse (Maybe Text)

An opaque string that indicates that the response contains only a subset of backups. Use this value in a subsequent DescribeBackups request to get more backups.

describeBackupsResponse_httpStatus :: Lens' DescribeBackupsResponse Int

The response's http status code.