Network.AWS.CloudHSMv2.Types

Description

Service Configuration

defaultService :: Service

API version 2017-04-28 of the Amazon CloudHSM V2 SDK configuration.

Errors

_CloudHsmInternalFailureException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because of an AWS CloudHSM internal failure. The request can be retried.

_CloudHsmResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because it refers to a resource that cannot be found.

_CloudHsmAccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because the requester does not have permission to perform the requested operation.

_CloudHsmInvalidRequestException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because it is not a valid request.

_CloudHsmServiceException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because an error occurred.

_CloudHsmTagException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because of a tagging failure. Verify the tag conditions in all applicable policies, and then retry the request.

BackupPolicy

newtype BackupPolicy

Constructors

BackupPolicy'

Fields

fromBackupPolicy :: Text

Bundled Patterns

pattern BackupPolicy_DEFAULT :: BackupPolicy

BackupRetentionType

newtype BackupRetentionType

Constructors

BackupRetentionType'

Fields

fromBackupRetentionType :: Text

Bundled Patterns

pattern BackupRetentionType_DAYS :: BackupRetentionType

BackupState

newtype BackupState

Constructors

BackupState'

Fields

fromBackupState :: Text

Bundled Patterns

pattern BackupState_CREATE_IN_PROGRESS :: BackupState

pattern BackupState_DELETED :: BackupState

pattern BackupState_PENDING_DELETION :: BackupState

pattern BackupState_READY :: BackupState

ClusterState

newtype ClusterState

Constructors

ClusterState'

Fields

fromClusterState :: Text

Bundled Patterns

pattern ClusterState_ACTIVE :: ClusterState

pattern ClusterState_CREATE_IN_PROGRESS :: ClusterState

pattern ClusterState_DEGRADED :: ClusterState

pattern ClusterState_DELETED :: ClusterState

pattern ClusterState_DELETE_IN_PROGRESS :: ClusterState

pattern ClusterState_INITIALIZED :: ClusterState

pattern ClusterState_INITIALIZE_IN_PROGRESS :: ClusterState

pattern ClusterState_UNINITIALIZED :: ClusterState

pattern ClusterState_UPDATE_IN_PROGRESS :: ClusterState

HsmState

newtype HsmState

Constructors

HsmState'

Fields

fromHsmState :: Text

Bundled Patterns

pattern HsmState_ACTIVE :: HsmState

pattern HsmState_CREATE_IN_PROGRESS :: HsmState

pattern HsmState_DEGRADED :: HsmState

pattern HsmState_DELETED :: HsmState

pattern HsmState_DELETE_IN_PROGRESS :: HsmState

Backup

data Backup

Contains information about a backup of an AWS CloudHSM cluster. All backup objects contain the BackupId, BackupState, ClusterId, and CreateTimestamp parameters. Backups that were copied into a destination region additionally contain the CopyTimestamp, SourceBackup, SourceCluster, and SourceRegion parameters. A backup that is pending deletion will include the DeleteTimestamp parameter.

See: newBackup smart constructor.

Constructors

Backup'

Fields

clusterId :: Maybe Text

The identifier (ID) of the cluster that was backed up.

backupState :: Maybe BackupState

The state of the backup.

sourceBackup :: Maybe Text

The identifier (ID) of the source backup from which the new backup was copied.

copyTimestamp :: Maybe POSIX

The date and time when the backup was copied from a source backup.

neverExpires :: Maybe Bool

Specifies whether the service should exempt a backup from the retention policy for the cluster. True exempts a backup from the retention policy. False means the service applies the backup retention policy defined at the cluster.

createTimestamp :: Maybe POSIX

The date and time when the backup was created.

deleteTimestamp :: Maybe POSIX

The date and time when the backup will be permanently deleted.

sourceCluster :: Maybe Text

The identifier (ID) of the cluster containing the source backup from which the new backup was copied.

tagList :: Maybe [Tag]

The list of tags for the backup.

sourceRegion :: Maybe Text

The AWS Region that contains the source backup from which the new backup was copied.

backupId :: Text

The identifier (ID) of the backup.

newBackup

Arguments

:: Text
$sel:backupId:Backup'

-> Backup

Create a value of Backup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clusterId:Backup', backup_clusterId - The identifier (ID) of the cluster that was backed up.

$sel:backupState:Backup', backup_backupState - The state of the backup.

$sel:sourceBackup:Backup', backup_sourceBackup - The identifier (ID) of the source backup from which the new backup was copied.

$sel:copyTimestamp:Backup', backup_copyTimestamp - The date and time when the backup was copied from a source backup.

$sel:neverExpires:Backup', backup_neverExpires - Specifies whether the service should exempt a backup from the retention policy for the cluster. True exempts a backup from the retention policy. False means the service applies the backup retention policy defined at the cluster.

$sel:createTimestamp:Backup', backup_createTimestamp - The date and time when the backup was created.

$sel:deleteTimestamp:Backup', backup_deleteTimestamp - The date and time when the backup will be permanently deleted.

$sel:sourceCluster:Backup', backup_sourceCluster - The identifier (ID) of the cluster containing the source backup from which the new backup was copied.

$sel:tagList:Backup', backup_tagList - The list of tags for the backup.

$sel:sourceRegion:Backup', backup_sourceRegion - The AWS Region that contains the source backup from which the new backup was copied.

$sel:backupId:Backup', backup_backupId - The identifier (ID) of the backup.

backup_clusterId :: Lens' Backup (Maybe Text)

The identifier (ID) of the cluster that was backed up.

backup_backupState :: Lens' Backup (Maybe BackupState)

The state of the backup.

backup_sourceBackup :: Lens' Backup (Maybe Text)

The identifier (ID) of the source backup from which the new backup was copied.

backup_copyTimestamp :: Lens' Backup (Maybe UTCTime)

The date and time when the backup was copied from a source backup.

backup_neverExpires :: Lens' Backup (Maybe Bool)

Specifies whether the service should exempt a backup from the retention policy for the cluster. True exempts a backup from the retention policy. False means the service applies the backup retention policy defined at the cluster.

backup_createTimestamp :: Lens' Backup (Maybe UTCTime)

The date and time when the backup was created.

backup_deleteTimestamp :: Lens' Backup (Maybe UTCTime)

The date and time when the backup will be permanently deleted.

backup_sourceCluster :: Lens' Backup (Maybe Text)

The identifier (ID) of the cluster containing the source backup from which the new backup was copied.

backup_tagList :: Lens' Backup (Maybe [Tag])

The list of tags for the backup.

backup_sourceRegion :: Lens' Backup (Maybe Text)

The AWS Region that contains the source backup from which the new backup was copied.

backup_backupId :: Lens' Backup Text

The identifier (ID) of the backup.

BackupRetentionPolicy

data BackupRetentionPolicy

A policy that defines the number of days to retain backups.

See: newBackupRetentionPolicy smart constructor.

Constructors

BackupRetentionPolicy'

Fields

value :: Maybe Text

Use a value between 7 - 379.

type' :: Maybe BackupRetentionType

The type of backup retention policy. For the DAYS type, the value is the number of days to retain backups.

newBackupRetentionPolicy :: BackupRetentionPolicy

Create a value of BackupRetentionPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:BackupRetentionPolicy', backupRetentionPolicy_value - Use a value between 7 - 379.

$sel:type':BackupRetentionPolicy', backupRetentionPolicy_type - The type of backup retention policy. For the DAYS type, the value is the number of days to retain backups.

backupRetentionPolicy_value :: Lens' BackupRetentionPolicy (Maybe Text)

Use a value between 7 - 379.

backupRetentionPolicy_type :: Lens' BackupRetentionPolicy (Maybe BackupRetentionType)

The type of backup retention policy. For the DAYS type, the value is the number of days to retain backups.

Certificates

data Certificates

Contains one or more certificates or a certificate signing request (CSR).

See: newCertificates smart constructor.

Constructors

Certificates'

Fields

awsHardwareCertificate :: Maybe Text

The HSM hardware certificate issued (signed) by AWS CloudHSM.

hsmCertificate :: Maybe Text

The HSM certificate issued (signed) by the HSM hardware.

clusterCsr :: Maybe Text

The cluster's certificate signing request (CSR). The CSR exists only when the cluster's state is UNINITIALIZED.

clusterCertificate :: Maybe Text

The cluster certificate issued (signed) by the issuing certificate authority (CA) of the cluster's owner.

manufacturerHardwareCertificate :: Maybe Text

The HSM hardware certificate issued (signed) by the hardware manufacturer.

newCertificates :: Certificates

Create a value of Certificates with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:awsHardwareCertificate:Certificates', certificates_awsHardwareCertificate - The HSM hardware certificate issued (signed) by AWS CloudHSM.

$sel:hsmCertificate:Certificates', certificates_hsmCertificate - The HSM certificate issued (signed) by the HSM hardware.

$sel:clusterCsr:Certificates', certificates_clusterCsr - The cluster's certificate signing request (CSR). The CSR exists only when the cluster's state is UNINITIALIZED.

$sel:clusterCertificate:Certificates', certificates_clusterCertificate - The cluster certificate issued (signed) by the issuing certificate authority (CA) of the cluster's owner.

$sel:manufacturerHardwareCertificate:Certificates', certificates_manufacturerHardwareCertificate - The HSM hardware certificate issued (signed) by the hardware manufacturer.

certificates_awsHardwareCertificate :: Lens' Certificates (Maybe Text)

The HSM hardware certificate issued (signed) by AWS CloudHSM.

certificates_hsmCertificate :: Lens' Certificates (Maybe Text)

The HSM certificate issued (signed) by the HSM hardware.

certificates_clusterCsr :: Lens' Certificates (Maybe Text)

The cluster's certificate signing request (CSR). The CSR exists only when the cluster's state is UNINITIALIZED.

certificates_clusterCertificate :: Lens' Certificates (Maybe Text)

The cluster certificate issued (signed) by the issuing certificate authority (CA) of the cluster's owner.

certificates_manufacturerHardwareCertificate :: Lens' Certificates (Maybe Text)

The HSM hardware certificate issued (signed) by the hardware manufacturer.

Cluster

data Cluster

Contains information about an AWS CloudHSM cluster.

See: newCluster smart constructor.

Constructors

Cluster'

Fields

clusterId :: Maybe Text

The cluster's identifier (ID).

stateMessage :: Maybe Text

A description of the cluster's state.

backupPolicy :: Maybe BackupPolicy

The cluster's backup policy.

createTimestamp :: Maybe POSIX

The date and time when the cluster was created.

subnetMapping :: Maybe (HashMap Text Text)

A map from availability zone to the cluster’s subnet in that availability zone.

state :: Maybe ClusterState

The cluster's state.

preCoPassword :: Maybe Text

The default password for the cluster's Pre-Crypto Officer (PRECO) user.

securityGroup :: Maybe Text

The identifier (ID) of the cluster's security group.

certificates :: Maybe Certificates

Contains one or more certificates or a certificate signing request (CSR).

sourceBackupId :: Maybe Text

The identifier (ID) of the backup used to create the cluster. This value exists only when the cluster was created from a backup.

hsmType :: Maybe Text

The type of HSM that the cluster contains.

tagList :: Maybe [Tag]

The list of tags for the cluster.

vpcId :: Maybe Text

The identifier (ID) of the virtual private cloud (VPC) that contains the cluster.

hsms :: Maybe [Hsm]

Contains information about the HSMs in the cluster.

backupRetentionPolicy :: Maybe BackupRetentionPolicy

A policy that defines how the service retains backups.

newCluster :: Cluster

Create a value of Cluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clusterId:Cluster', cluster_clusterId - The cluster's identifier (ID).

$sel:stateMessage:Cluster', cluster_stateMessage - A description of the cluster's state.

$sel:backupPolicy:Cluster', cluster_backupPolicy - The cluster's backup policy.

$sel:createTimestamp:Cluster', cluster_createTimestamp - The date and time when the cluster was created.

$sel:subnetMapping:Cluster', cluster_subnetMapping - A map from availability zone to the cluster’s subnet in that availability zone.

$sel:state:Cluster', cluster_state - The cluster's state.

$sel:preCoPassword:Cluster', cluster_preCoPassword - The default password for the cluster's Pre-Crypto Officer (PRECO) user.

$sel:securityGroup:Cluster', cluster_securityGroup - The identifier (ID) of the cluster's security group.

$sel:certificates:Cluster', cluster_certificates - Contains one or more certificates or a certificate signing request (CSR).

$sel:sourceBackupId:Cluster', cluster_sourceBackupId - The identifier (ID) of the backup used to create the cluster. This value exists only when the cluster was created from a backup.

$sel:hsmType:Cluster', cluster_hsmType - The type of HSM that the cluster contains.

$sel:tagList:Cluster', cluster_tagList - The list of tags for the cluster.

$sel:vpcId:Cluster', cluster_vpcId - The identifier (ID) of the virtual private cloud (VPC) that contains the cluster.

$sel:hsms:Cluster', cluster_hsms - Contains information about the HSMs in the cluster.

$sel:backupRetentionPolicy:Cluster', cluster_backupRetentionPolicy - A policy that defines how the service retains backups.

cluster_clusterId :: Lens' Cluster (Maybe Text)

The cluster's identifier (ID).

cluster_stateMessage :: Lens' Cluster (Maybe Text)

A description of the cluster's state.

cluster_backupPolicy :: Lens' Cluster (Maybe BackupPolicy)

The cluster's backup policy.

cluster_createTimestamp :: Lens' Cluster (Maybe UTCTime)

The date and time when the cluster was created.

cluster_subnetMapping :: Lens' Cluster (Maybe (HashMap Text Text))

A map from availability zone to the cluster’s subnet in that availability zone.

cluster_state :: Lens' Cluster (Maybe ClusterState)

The cluster's state.

cluster_preCoPassword :: Lens' Cluster (Maybe Text)

The default password for the cluster's Pre-Crypto Officer (PRECO) user.

cluster_securityGroup :: Lens' Cluster (Maybe Text)

The identifier (ID) of the cluster's security group.

cluster_certificates :: Lens' Cluster (Maybe Certificates)

Contains one or more certificates or a certificate signing request (CSR).

cluster_sourceBackupId :: Lens' Cluster (Maybe Text)

The identifier (ID) of the backup used to create the cluster. This value exists only when the cluster was created from a backup.

cluster_hsmType :: Lens' Cluster (Maybe Text)

The type of HSM that the cluster contains.

cluster_tagList :: Lens' Cluster (Maybe [Tag])

The list of tags for the cluster.

cluster_vpcId :: Lens' Cluster (Maybe Text)

The identifier (ID) of the virtual private cloud (VPC) that contains the cluster.

cluster_hsms :: Lens' Cluster (Maybe [Hsm])

Contains information about the HSMs in the cluster.

cluster_backupRetentionPolicy :: Lens' Cluster (Maybe BackupRetentionPolicy)

A policy that defines how the service retains backups.

DestinationBackup

data DestinationBackup

Contains information about the backup that will be copied and created by the CopyBackupToRegion operation.

See: newDestinationBackup smart constructor.

Constructors

DestinationBackup'

Fields

sourceBackup :: Maybe Text

The identifier (ID) of the source backup from which the new backup was copied.

createTimestamp :: Maybe POSIX

The date and time when both the source backup was created.

sourceCluster :: Maybe Text

The identifier (ID) of the cluster containing the source backup from which the new backup was copied.

sourceRegion :: Maybe Text

The AWS region that contains the source backup from which the new backup was copied.

newDestinationBackup :: DestinationBackup

Create a value of DestinationBackup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceBackup:DestinationBackup', destinationBackup_sourceBackup - The identifier (ID) of the source backup from which the new backup was copied.

$sel:createTimestamp:DestinationBackup', destinationBackup_createTimestamp - The date and time when both the source backup was created.

$sel:sourceCluster:DestinationBackup', destinationBackup_sourceCluster - The identifier (ID) of the cluster containing the source backup from which the new backup was copied.

$sel:sourceRegion:DestinationBackup', destinationBackup_sourceRegion - The AWS region that contains the source backup from which the new backup was copied.

destinationBackup_sourceBackup :: Lens' DestinationBackup (Maybe Text)

The identifier (ID) of the source backup from which the new backup was copied.

destinationBackup_createTimestamp :: Lens' DestinationBackup (Maybe UTCTime)

The date and time when both the source backup was created.

destinationBackup_sourceCluster :: Lens' DestinationBackup (Maybe Text)

The identifier (ID) of the cluster containing the source backup from which the new backup was copied.

destinationBackup_sourceRegion :: Lens' DestinationBackup (Maybe Text)

The AWS region that contains the source backup from which the new backup was copied.

Hsm

data Hsm

Contains information about a hardware security module (HSM) in an AWS CloudHSM cluster.

See: newHsm smart constructor.

Constructors

Hsm'

Fields

clusterId :: Maybe Text

The identifier (ID) of the cluster that contains the HSM.

stateMessage :: Maybe Text

A description of the HSM's state.

eniIp :: Maybe Text

The IP address of the HSM's elastic network interface (ENI).

eniId :: Maybe Text

The identifier (ID) of the HSM's elastic network interface (ENI).

availabilityZone :: Maybe Text

The Availability Zone that contains the HSM.

state :: Maybe HsmState

The HSM's state.

subnetId :: Maybe Text

The subnet that contains the HSM's elastic network interface (ENI).

hsmId :: Text

The HSM's identifier (ID).

newHsm

Arguments

:: Text
$sel:hsmId:Hsm'

-> Hsm

Create a value of Hsm with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clusterId:Hsm', hsm_clusterId - The identifier (ID) of the cluster that contains the HSM.

$sel:stateMessage:Hsm', hsm_stateMessage - A description of the HSM's state.

$sel:eniIp:Hsm', hsm_eniIp - The IP address of the HSM's elastic network interface (ENI).

$sel:eniId:Hsm', hsm_eniId - The identifier (ID) of the HSM's elastic network interface (ENI).

$sel:availabilityZone:Hsm', hsm_availabilityZone - The Availability Zone that contains the HSM.

$sel:state:Hsm', hsm_state - The HSM's state.

$sel:subnetId:Hsm', hsm_subnetId - The subnet that contains the HSM's elastic network interface (ENI).

$sel:hsmId:Hsm', hsm_hsmId - The HSM's identifier (ID).

hsm_clusterId :: Lens' Hsm (Maybe Text)

The identifier (ID) of the cluster that contains the HSM.

hsm_stateMessage :: Lens' Hsm (Maybe Text)

A description of the HSM's state.

hsm_eniIp :: Lens' Hsm (Maybe Text)

The IP address of the HSM's elastic network interface (ENI).

hsm_eniId :: Lens' Hsm (Maybe Text)

The identifier (ID) of the HSM's elastic network interface (ENI).

hsm_availabilityZone :: Lens' Hsm (Maybe Text)

The Availability Zone that contains the HSM.

hsm_state :: Lens' Hsm (Maybe HsmState)

The HSM's state.

hsm_subnetId :: Lens' Hsm (Maybe Text)

The subnet that contains the HSM's elastic network interface (ENI).

hsm_hsmId :: Lens' Hsm Text

The HSM's identifier (ID).

Tag

data Tag

Contains a tag. A tag is a key-value pair.

See: newTag smart constructor.

Constructors

Tag'

Fields

key :: Text

The key of the tag.

value :: Text

The value of the tag.

newTag

Arguments

:: Text
$sel:key:Tag'

-> Text
$sel:value:Tag'

-> Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:Tag', tag_key - The key of the tag.

$sel:value:Tag', tag_value - The value of the tag.

tag_key :: Lens' Tag Text

The key of the tag.

tag_value :: Lens' Tag Text

The value of the tag.