Network.AWS.CloudHSMv2.Types.Backup

Description

Documentation

data Backup

Contains information about a backup of an AWS CloudHSM cluster. All backup objects contain the BackupId, BackupState, ClusterId, and CreateTimestamp parameters. Backups that were copied into a destination region additionally contain the CopyTimestamp, SourceBackup, SourceCluster, and SourceRegion parameters. A backup that is pending deletion will include the DeleteTimestamp parameter.

See: newBackup smart constructor.

Constructors

Backup'

Fields

clusterId :: Maybe Text

The identifier (ID) of the cluster that was backed up.

backupState :: Maybe BackupState

The state of the backup.

sourceBackup :: Maybe Text

The identifier (ID) of the source backup from which the new backup was copied.

copyTimestamp :: Maybe POSIX

The date and time when the backup was copied from a source backup.

neverExpires :: Maybe Bool

Specifies whether the service should exempt a backup from the retention policy for the cluster. True exempts a backup from the retention policy. False means the service applies the backup retention policy defined at the cluster.

createTimestamp :: Maybe POSIX

The date and time when the backup was created.

deleteTimestamp :: Maybe POSIX

The date and time when the backup will be permanently deleted.

sourceCluster :: Maybe Text

The identifier (ID) of the cluster containing the source backup from which the new backup was copied.

tagList :: Maybe [Tag]

The list of tags for the backup.

sourceRegion :: Maybe Text

The AWS Region that contains the source backup from which the new backup was copied.

backupId :: Text

The identifier (ID) of the backup.

newBackup

Arguments

:: Text
$sel:backupId:Backup'

-> Backup

Create a value of Backup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clusterId:Backup', backup_clusterId - The identifier (ID) of the cluster that was backed up.

$sel:backupState:Backup', backup_backupState - The state of the backup.

$sel:sourceBackup:Backup', backup_sourceBackup - The identifier (ID) of the source backup from which the new backup was copied.

$sel:copyTimestamp:Backup', backup_copyTimestamp - The date and time when the backup was copied from a source backup.

$sel:neverExpires:Backup', backup_neverExpires - Specifies whether the service should exempt a backup from the retention policy for the cluster. True exempts a backup from the retention policy. False means the service applies the backup retention policy defined at the cluster.

$sel:createTimestamp:Backup', backup_createTimestamp - The date and time when the backup was created.

$sel:deleteTimestamp:Backup', backup_deleteTimestamp - The date and time when the backup will be permanently deleted.

$sel:sourceCluster:Backup', backup_sourceCluster - The identifier (ID) of the cluster containing the source backup from which the new backup was copied.

$sel:tagList:Backup', backup_tagList - The list of tags for the backup.

$sel:sourceRegion:Backup', backup_sourceRegion - The AWS Region that contains the source backup from which the new backup was copied.

$sel:backupId:Backup', backup_backupId - The identifier (ID) of the backup.

backup_clusterId :: Lens' Backup (Maybe Text)

The identifier (ID) of the cluster that was backed up.

backup_backupState :: Lens' Backup (Maybe BackupState)

The state of the backup.

backup_sourceBackup :: Lens' Backup (Maybe Text)

The identifier (ID) of the source backup from which the new backup was copied.

backup_copyTimestamp :: Lens' Backup (Maybe UTCTime)

The date and time when the backup was copied from a source backup.

backup_neverExpires :: Lens' Backup (Maybe Bool)

Specifies whether the service should exempt a backup from the retention policy for the cluster. True exempts a backup from the retention policy. False means the service applies the backup retention policy defined at the cluster.

backup_createTimestamp :: Lens' Backup (Maybe UTCTime)

The date and time when the backup was created.

backup_deleteTimestamp :: Lens' Backup (Maybe UTCTime)

The date and time when the backup will be permanently deleted.

backup_sourceCluster :: Lens' Backup (Maybe Text)

The identifier (ID) of the cluster containing the source backup from which the new backup was copied.

backup_tagList :: Lens' Backup (Maybe [Tag])

The list of tags for the backup.

backup_sourceRegion :: Lens' Backup (Maybe Text)

The AWS Region that contains the source backup from which the new backup was copied.

backup_backupId :: Lens' Backup Text

The identifier (ID) of the backup.