Network.AWS.CloudHSMv2.Types.Cluster

Description

Documentation

data Cluster

Contains information about an AWS CloudHSM cluster.

See: newCluster smart constructor.

Constructors

Cluster'

Fields

preCoPassword :: Maybe Text

The default password for the cluster's Pre-Crypto Officer (PRECO) user.

stateMessage :: Maybe Text

A description of the cluster's state.

state :: Maybe ClusterState

The cluster's state.

subnetMapping :: Maybe (HashMap Text Text)

A map from availability zone to the cluster’s subnet in that availability zone.

backupRetentionPolicy :: Maybe BackupRetentionPolicy

A policy that defines how the service retains backups.

hsms :: Maybe [Hsm]

Contains information about the HSMs in the cluster.

vpcId :: Maybe Text

The identifier (ID) of the virtual private cloud (VPC) that contains the cluster.

tagList :: Maybe [Tag]

The list of tags for the cluster.

sourceBackupId :: Maybe Text

The identifier (ID) of the backup used to create the cluster. This value exists only when the cluster was created from a backup.

certificates :: Maybe Certificates

Contains one or more certificates or a certificate signing request (CSR).

securityGroup :: Maybe Text

The identifier (ID) of the cluster's security group.

clusterId :: Maybe Text

The cluster's identifier (ID).

createTimestamp :: Maybe POSIX

The date and time when the cluster was created.

backupPolicy :: Maybe BackupPolicy

The cluster's backup policy.

hsmType :: Maybe Text

The type of HSM that the cluster contains.

newCluster :: Cluster

Create a value of Cluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:preCoPassword:Cluster', cluster_preCoPassword - The default password for the cluster's Pre-Crypto Officer (PRECO) user.

$sel:stateMessage:Cluster', cluster_stateMessage - A description of the cluster's state.

$sel:state:Cluster', cluster_state - The cluster's state.

$sel:subnetMapping:Cluster', cluster_subnetMapping - A map from availability zone to the cluster’s subnet in that availability zone.

$sel:backupRetentionPolicy:Cluster', cluster_backupRetentionPolicy - A policy that defines how the service retains backups.

$sel:hsms:Cluster', cluster_hsms - Contains information about the HSMs in the cluster.

$sel:vpcId:Cluster', cluster_vpcId - The identifier (ID) of the virtual private cloud (VPC) that contains the cluster.

$sel:tagList:Cluster', cluster_tagList - The list of tags for the cluster.

$sel:sourceBackupId:Cluster', cluster_sourceBackupId - The identifier (ID) of the backup used to create the cluster. This value exists only when the cluster was created from a backup.

$sel:certificates:Cluster', cluster_certificates - Contains one or more certificates or a certificate signing request (CSR).

$sel:securityGroup:Cluster', cluster_securityGroup - The identifier (ID) of the cluster's security group.

$sel:clusterId:Cluster', cluster_clusterId - The cluster's identifier (ID).

$sel:createTimestamp:Cluster', cluster_createTimestamp - The date and time when the cluster was created.

$sel:backupPolicy:Cluster', cluster_backupPolicy - The cluster's backup policy.

$sel:hsmType:Cluster', cluster_hsmType - The type of HSM that the cluster contains.

cluster_preCoPassword :: Lens' Cluster (Maybe Text)

The default password for the cluster's Pre-Crypto Officer (PRECO) user.

cluster_stateMessage :: Lens' Cluster (Maybe Text)

A description of the cluster's state.

cluster_state :: Lens' Cluster (Maybe ClusterState)

The cluster's state.

cluster_subnetMapping :: Lens' Cluster (Maybe (HashMap Text Text))

A map from availability zone to the cluster’s subnet in that availability zone.

cluster_backupRetentionPolicy :: Lens' Cluster (Maybe BackupRetentionPolicy)

A policy that defines how the service retains backups.

cluster_hsms :: Lens' Cluster (Maybe [Hsm])

Contains information about the HSMs in the cluster.

cluster_vpcId :: Lens' Cluster (Maybe Text)

The identifier (ID) of the virtual private cloud (VPC) that contains the cluster.

cluster_tagList :: Lens' Cluster (Maybe [Tag])

The list of tags for the cluster.

cluster_sourceBackupId :: Lens' Cluster (Maybe Text)

The identifier (ID) of the backup used to create the cluster. This value exists only when the cluster was created from a backup.

cluster_certificates :: Lens' Cluster (Maybe Certificates)

Contains one or more certificates or a certificate signing request (CSR).

cluster_securityGroup :: Lens' Cluster (Maybe Text)

The identifier (ID) of the cluster's security group.

cluster_clusterId :: Lens' Cluster (Maybe Text)

The cluster's identifier (ID).

cluster_createTimestamp :: Lens' Cluster (Maybe UTCTime)

The date and time when the cluster was created.

cluster_backupPolicy :: Lens' Cluster (Maybe BackupPolicy)

The cluster's backup policy.

cluster_hsmType :: Lens' Cluster (Maybe Text)

The type of HSM that the cluster contains.