Network.AWS.CloudHSMv2.Types.Hsm

Description

Documentation

data Hsm

Contains information about a hardware security module (HSM) in an AWS CloudHSM cluster.

See: newHsm smart constructor.

Constructors

Hsm'

Fields

stateMessage :: Maybe Text

A description of the HSM's state.

state :: Maybe HsmState

The HSM's state.

eniId :: Maybe Text

The identifier (ID) of the HSM's elastic network interface (ENI).

subnetId :: Maybe Text

The subnet that contains the HSM's elastic network interface (ENI).

availabilityZone :: Maybe Text

The Availability Zone that contains the HSM.

clusterId :: Maybe Text

The identifier (ID) of the cluster that contains the HSM.

eniIp :: Maybe Text

The IP address of the HSM's elastic network interface (ENI).

hsmId :: Text

The HSM's identifier (ID).

newHsm

Arguments

:: Text
$sel:hsmId:Hsm'

-> Hsm

Create a value of Hsm with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stateMessage:Hsm', hsm_stateMessage - A description of the HSM's state.

$sel:state:Hsm', hsm_state - The HSM's state.

$sel:eniId:Hsm', hsm_eniId - The identifier (ID) of the HSM's elastic network interface (ENI).

$sel:subnetId:Hsm', hsm_subnetId - The subnet that contains the HSM's elastic network interface (ENI).

$sel:availabilityZone:Hsm', hsm_availabilityZone - The Availability Zone that contains the HSM.

$sel:clusterId:Hsm', hsm_clusterId - The identifier (ID) of the cluster that contains the HSM.

$sel:eniIp:Hsm', hsm_eniIp - The IP address of the HSM's elastic network interface (ENI).

$sel:hsmId:Hsm', hsm_hsmId - The HSM's identifier (ID).

hsm_stateMessage :: Lens' Hsm (Maybe Text)

A description of the HSM's state.

hsm_state :: Lens' Hsm (Maybe HsmState)

The HSM's state.

hsm_eniId :: Lens' Hsm (Maybe Text)

The identifier (ID) of the HSM's elastic network interface (ENI).

hsm_subnetId :: Lens' Hsm (Maybe Text)

The subnet that contains the HSM's elastic network interface (ENI).

hsm_availabilityZone :: Lens' Hsm (Maybe Text)

The Availability Zone that contains the HSM.

hsm_clusterId :: Lens' Hsm (Maybe Text)

The identifier (ID) of the cluster that contains the HSM.

hsm_eniIp :: Lens' Hsm (Maybe Text)

The IP address of the HSM's elastic network interface (ENI).

hsm_hsmId :: Lens' Hsm Text

The HSM's identifier (ID).