Network.AWS.CloudSearch.BuildSuggesters

Description

Indexes the search suggestions. For more information, see Configuring Suggesters in the Amazon CloudSearch Developer Guide.

Creating a Request

data BuildSuggesters

Container for the parameters to the BuildSuggester operation. Specifies the name of the domain you want to update.

See: newBuildSuggesters smart constructor.

Constructors

BuildSuggesters'

Fields

domainName :: Text

newBuildSuggesters

Arguments

:: Text
$sel:domainName:BuildSuggesters'

-> BuildSuggesters

Create a value of BuildSuggesters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainName:BuildSuggesters', buildSuggesters_domainName - Undocumented member.

Request Lenses

buildSuggesters_domainName :: Lens' BuildSuggesters Text

Undocumented member.

Destructuring the Response

data BuildSuggestersResponse

The result of a BuildSuggester request. Contains a list of the fields used for suggestions.

See: newBuildSuggestersResponse smart constructor.

Constructors

BuildSuggestersResponse'

Fields

fieldNames :: Maybe [Text]

httpStatus :: Int

The response's http status code.

newBuildSuggestersResponse

Arguments

:: Int
$sel:httpStatus:BuildSuggestersResponse'

-> BuildSuggestersResponse

Create a value of BuildSuggestersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fieldNames:BuildSuggestersResponse', buildSuggestersResponse_fieldNames - Undocumented member.

$sel:httpStatus:BuildSuggestersResponse', buildSuggestersResponse_httpStatus - The response's http status code.

Response Lenses

buildSuggestersResponse_fieldNames :: Lens' BuildSuggestersResponse (Maybe [Text])

Undocumented member.

buildSuggestersResponse_httpStatus :: Lens' BuildSuggestersResponse Int

The response's http status code.