Network.AWS.CloudSearch.DefineIndexField

Description

Configures an IndexField for the search domain. Used to create new fields and modify existing ones. You must specify the name of the domain you are configuring and an index field configuration. The index field configuration specifies a unique name, the index field type, and the options you want to configure for the field. The options you can specify depend on the IndexFieldType. If the field exists, the new configuration replaces the old one. For more information, see Configuring Index Fields in the Amazon CloudSearch Developer Guide.

Creating a Request

data DefineIndexField

Container for the parameters to the DefineIndexField operation. Specifies the name of the domain you want to update and the index field configuration.

See: newDefineIndexField smart constructor.

Constructors

DefineIndexField'

Fields

domainName :: Text

indexField :: IndexField

The index field and field options you want to configure.

newDefineIndexField

Arguments

:: Text
$sel:domainName:DefineIndexField'

-> IndexField
$sel:indexField:DefineIndexField'

-> DefineIndexField

Create a value of DefineIndexField with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainName:DefineIndexField', defineIndexField_domainName - Undocumented member.

$sel:indexField:DefineIndexField', defineIndexField_indexField - The index field and field options you want to configure.

Request Lenses

defineIndexField_domainName :: Lens' DefineIndexField Text

Undocumented member.

defineIndexField_indexField :: Lens' DefineIndexField IndexField

The index field and field options you want to configure.

Destructuring the Response

data DefineIndexFieldResponse

The result of a DefineIndexField request. Contains the status of the newly-configured index field.

See: newDefineIndexFieldResponse smart constructor.

Constructors

DefineIndexFieldResponse'

Fields

httpStatus :: Int

The response's http status code.

indexField :: IndexFieldStatus

newDefineIndexFieldResponse

Arguments

:: Int
$sel:httpStatus:DefineIndexFieldResponse'

-> IndexFieldStatus
$sel:indexField:DefineIndexField'

-> DefineIndexFieldResponse

Create a value of DefineIndexFieldResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DefineIndexFieldResponse', defineIndexFieldResponse_httpStatus - The response's http status code.

$sel:indexField:DefineIndexField', defineIndexFieldResponse_indexField - Undocumented member.

Response Lenses

defineIndexFieldResponse_httpStatus :: Lens' DefineIndexFieldResponse Int

The response's http status code.

defineIndexFieldResponse_indexField :: Lens' DefineIndexFieldResponse IndexFieldStatus

Undocumented member.