Network.AWS.CloudSearch.DefineSuggester

Description

Configures a suggester for a domain. A suggester enables you to display possible matches before users finish typing their queries. When you configure a suggester, you must specify the name of the text field you want to search for possible matches and a unique name for the suggester. For more information, see Getting Search Suggestions in the Amazon CloudSearch Developer Guide.

Creating a Request

data DefineSuggester

Container for the parameters to the DefineSuggester operation. Specifies the name of the domain you want to update and the suggester configuration.

See: newDefineSuggester smart constructor.

Constructors

DefineSuggester'

Fields

domainName :: Text

suggester :: Suggester

newDefineSuggester

Arguments

:: Text
$sel:domainName:DefineSuggester'

-> Suggester
$sel:suggester:DefineSuggester'

-> DefineSuggester

Create a value of DefineSuggester with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainName:DefineSuggester', defineSuggester_domainName - Undocumented member.

$sel:suggester:DefineSuggester', defineSuggester_suggester - Undocumented member.

Request Lenses

defineSuggester_domainName :: Lens' DefineSuggester Text

Undocumented member.

defineSuggester_suggester :: Lens' DefineSuggester Suggester

Undocumented member.

Destructuring the Response

data DefineSuggesterResponse

The result of a DefineSuggester request. Contains the status of the newly-configured suggester.

See: newDefineSuggesterResponse smart constructor.

Constructors

DefineSuggesterResponse'

Fields

httpStatus :: Int

The response's http status code.

suggester :: SuggesterStatus

newDefineSuggesterResponse

Arguments

:: Int
$sel:httpStatus:DefineSuggesterResponse'

-> SuggesterStatus
$sel:suggester:DefineSuggester'

-> DefineSuggesterResponse

Create a value of DefineSuggesterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DefineSuggesterResponse', defineSuggesterResponse_httpStatus - The response's http status code.

$sel:suggester:DefineSuggester', defineSuggesterResponse_suggester - Undocumented member.

Response Lenses

defineSuggesterResponse_httpStatus :: Lens' DefineSuggesterResponse Int

The response's http status code.

defineSuggesterResponse_suggester :: Lens' DefineSuggesterResponse SuggesterStatus

Undocumented member.