Network.AWS.CloudSearch.DeleteAnalysisScheme

Description

Deletes an analysis scheme. For more information, see Configuring Analysis Schemes in the Amazon CloudSearch Developer Guide.

Creating a Request

data DeleteAnalysisScheme

Container for the parameters to the DeleteAnalysisScheme operation. Specifies the name of the domain you want to update and the analysis scheme you want to delete.

See: newDeleteAnalysisScheme smart constructor.

Constructors

DeleteAnalysisScheme'

Fields

domainName :: Text

analysisSchemeName :: Text

The name of the analysis scheme you want to delete.

newDeleteAnalysisScheme

Arguments

:: Text
$sel:domainName:DeleteAnalysisScheme'

-> Text
$sel:analysisSchemeName:DeleteAnalysisScheme'

-> DeleteAnalysisScheme

Create a value of DeleteAnalysisScheme with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainName:DeleteAnalysisScheme', deleteAnalysisScheme_domainName - Undocumented member.

$sel:analysisSchemeName:DeleteAnalysisScheme', deleteAnalysisScheme_analysisSchemeName - The name of the analysis scheme you want to delete.

Request Lenses

deleteAnalysisScheme_domainName :: Lens' DeleteAnalysisScheme Text

Undocumented member.

deleteAnalysisScheme_analysisSchemeName :: Lens' DeleteAnalysisScheme Text

The name of the analysis scheme you want to delete.

Destructuring the Response

data DeleteAnalysisSchemeResponse

The result of a DeleteAnalysisScheme request. Contains the status of the deleted analysis scheme.

See: newDeleteAnalysisSchemeResponse smart constructor.

Constructors

DeleteAnalysisSchemeResponse'

Fields

httpStatus :: Int

The response's http status code.

analysisScheme :: AnalysisSchemeStatus

The status of the analysis scheme being deleted.

newDeleteAnalysisSchemeResponse

Arguments

:: Int
$sel:httpStatus:DeleteAnalysisSchemeResponse'

-> AnalysisSchemeStatus
$sel:analysisScheme:DeleteAnalysisSchemeResponse'

-> DeleteAnalysisSchemeResponse

Create a value of DeleteAnalysisSchemeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteAnalysisSchemeResponse', deleteAnalysisSchemeResponse_httpStatus - The response's http status code.

$sel:analysisScheme:DeleteAnalysisSchemeResponse', deleteAnalysisSchemeResponse_analysisScheme - The status of the analysis scheme being deleted.

Response Lenses

deleteAnalysisSchemeResponse_httpStatus :: Lens' DeleteAnalysisSchemeResponse Int

The response's http status code.

deleteAnalysisSchemeResponse_analysisScheme :: Lens' DeleteAnalysisSchemeResponse AnalysisSchemeStatus

The status of the analysis scheme being deleted.