Network.AWS.CloudSearch.DeleteSuggester

Description

Deletes a suggester. For more information, see Getting Search Suggestions in the Amazon CloudSearch Developer Guide.

Creating a Request

data DeleteSuggester

Container for the parameters to the DeleteSuggester operation. Specifies the name of the domain you want to update and name of the suggester you want to delete.

See: newDeleteSuggester smart constructor.

Constructors

DeleteSuggester'

Fields

domainName :: Text

suggesterName :: Text

Specifies the name of the suggester you want to delete.

newDeleteSuggester

Arguments

:: Text
$sel:domainName:DeleteSuggester'

-> Text
$sel:suggesterName:DeleteSuggester'

-> DeleteSuggester

Create a value of DeleteSuggester with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainName:DeleteSuggester', deleteSuggester_domainName - Undocumented member.

$sel:suggesterName:DeleteSuggester', deleteSuggester_suggesterName - Specifies the name of the suggester you want to delete.

Request Lenses

deleteSuggester_domainName :: Lens' DeleteSuggester Text

Undocumented member.

deleteSuggester_suggesterName :: Lens' DeleteSuggester Text

Specifies the name of the suggester you want to delete.

Destructuring the Response

data DeleteSuggesterResponse

The result of a DeleteSuggester request. Contains the status of the deleted suggester.

See: newDeleteSuggesterResponse smart constructor.

Constructors

DeleteSuggesterResponse'

Fields

httpStatus :: Int

The response's http status code.

suggester :: SuggesterStatus

The status of the suggester being deleted.

newDeleteSuggesterResponse

Arguments

:: Int
$sel:httpStatus:DeleteSuggesterResponse'

-> SuggesterStatus
$sel:suggester:DeleteSuggesterResponse'

-> DeleteSuggesterResponse

Create a value of DeleteSuggesterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteSuggesterResponse', deleteSuggesterResponse_httpStatus - The response's http status code.

$sel:suggester:DeleteSuggesterResponse', deleteSuggesterResponse_suggester - The status of the suggester being deleted.

Response Lenses

deleteSuggesterResponse_httpStatus :: Lens' DeleteSuggesterResponse Int

The response's http status code.

deleteSuggesterResponse_suggester :: Lens' DeleteSuggesterResponse SuggesterStatus

The status of the suggester being deleted.