Network.AWS.CloudSearch.DescribeAnalysisSchemes

Description

Gets the analysis schemes configured for a domain. An analysis scheme defines language-specific text processing options for a text field. Can be limited to specific analysis schemes by name. By default, shows all analysis schemes and includes any pending changes to the configuration. Set the Deployed option to true to show the active configuration and exclude pending changes. For more information, see Configuring Analysis Schemes in the Amazon CloudSearch Developer Guide.

Creating a Request

data DescribeAnalysisSchemes

Container for the parameters to the DescribeAnalysisSchemes operation. Specifies the name of the domain you want to describe. To limit the response to particular analysis schemes, specify the names of the analysis schemes you want to describe. To show the active configuration and exclude any pending changes, set the Deployed option to true.

See: newDescribeAnalysisSchemes smart constructor.

Constructors

DescribeAnalysisSchemes'

Fields

deployed :: Maybe Bool

Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

analysisSchemeNames :: Maybe [Text]

The analysis schemes you want to describe.

domainName :: Text

The name of the domain you want to describe.

newDescribeAnalysisSchemes

Arguments

:: Text
$sel:domainName:DescribeAnalysisSchemes'

-> DescribeAnalysisSchemes

Create a value of DescribeAnalysisSchemes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deployed:DescribeAnalysisSchemes', describeAnalysisSchemes_deployed - Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

$sel:analysisSchemeNames:DescribeAnalysisSchemes', describeAnalysisSchemes_analysisSchemeNames - The analysis schemes you want to describe.

$sel:domainName:DescribeAnalysisSchemes', describeAnalysisSchemes_domainName - The name of the domain you want to describe.

Request Lenses

describeAnalysisSchemes_deployed :: Lens' DescribeAnalysisSchemes (Maybe Bool)

Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

describeAnalysisSchemes_analysisSchemeNames :: Lens' DescribeAnalysisSchemes (Maybe [Text])

The analysis schemes you want to describe.

describeAnalysisSchemes_domainName :: Lens' DescribeAnalysisSchemes Text

The name of the domain you want to describe.

Destructuring the Response

data DescribeAnalysisSchemesResponse

The result of a DescribeAnalysisSchemes request. Contains the analysis schemes configured for the domain specified in the request.

See: newDescribeAnalysisSchemesResponse smart constructor.

Constructors

DescribeAnalysisSchemesResponse'

Fields

httpStatus :: Int

The response's http status code.

analysisSchemes :: [AnalysisSchemeStatus]

The analysis scheme descriptions.

newDescribeAnalysisSchemesResponse

Arguments

:: Int
$sel:httpStatus:DescribeAnalysisSchemesResponse'

-> DescribeAnalysisSchemesResponse

Create a value of DescribeAnalysisSchemesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DescribeAnalysisSchemesResponse', describeAnalysisSchemesResponse_httpStatus - The response's http status code.

$sel:analysisSchemes:DescribeAnalysisSchemesResponse', describeAnalysisSchemesResponse_analysisSchemes - The analysis scheme descriptions.

Response Lenses

describeAnalysisSchemesResponse_httpStatus :: Lens' DescribeAnalysisSchemesResponse Int

The response's http status code.

describeAnalysisSchemesResponse_analysisSchemes :: Lens' DescribeAnalysisSchemesResponse [AnalysisSchemeStatus]

The analysis scheme descriptions.