Network.AWS.CloudSearch.DescribeAvailabilityOptions

Description

Gets the availability options configured for a domain. By default, shows the configuration with any pending changes. Set the Deployed option to true to show the active configuration and exclude pending changes. For more information, see Configuring Availability Options in the Amazon CloudSearch Developer Guide.

Creating a Request

data DescribeAvailabilityOptions

Container for the parameters to the DescribeAvailabilityOptions operation. Specifies the name of the domain you want to describe. To show the active configuration and exclude any pending changes, set the Deployed option to true.

See: newDescribeAvailabilityOptions smart constructor.

Constructors

DescribeAvailabilityOptions'

Fields

deployed :: Maybe Bool

Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

domainName :: Text

The name of the domain you want to describe.

newDescribeAvailabilityOptions

Arguments

:: Text
$sel:domainName:DescribeAvailabilityOptions'

-> DescribeAvailabilityOptions

Create a value of DescribeAvailabilityOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deployed:DescribeAvailabilityOptions', describeAvailabilityOptions_deployed - Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

$sel:domainName:DescribeAvailabilityOptions', describeAvailabilityOptions_domainName - The name of the domain you want to describe.

Request Lenses

describeAvailabilityOptions_deployed :: Lens' DescribeAvailabilityOptions (Maybe Bool)

Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

describeAvailabilityOptions_domainName :: Lens' DescribeAvailabilityOptions Text

The name of the domain you want to describe.

Destructuring the Response

data DescribeAvailabilityOptionsResponse

The result of a DescribeAvailabilityOptions request. Indicates whether or not the Multi-AZ option is enabled for the domain specified in the request.

See: newDescribeAvailabilityOptionsResponse smart constructor.

Constructors

DescribeAvailabilityOptionsResponse'

Fields

availabilityOptions :: Maybe AvailabilityOptionsStatus

The availability options configured for the domain. Indicates whether Multi-AZ is enabled for the domain.

httpStatus :: Int

The response's http status code.

newDescribeAvailabilityOptionsResponse

Arguments

:: Int
$sel:httpStatus:DescribeAvailabilityOptionsResponse'

-> DescribeAvailabilityOptionsResponse

Create a value of DescribeAvailabilityOptionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:availabilityOptions:DescribeAvailabilityOptionsResponse', describeAvailabilityOptionsResponse_availabilityOptions - The availability options configured for the domain. Indicates whether Multi-AZ is enabled for the domain.

$sel:httpStatus:DescribeAvailabilityOptionsResponse', describeAvailabilityOptionsResponse_httpStatus - The response's http status code.

Response Lenses

describeAvailabilityOptionsResponse_availabilityOptions :: Lens' DescribeAvailabilityOptionsResponse (Maybe AvailabilityOptionsStatus)

The availability options configured for the domain. Indicates whether Multi-AZ is enabled for the domain.

describeAvailabilityOptionsResponse_httpStatus :: Lens' DescribeAvailabilityOptionsResponse Int

The response's http status code.