Network.AWS.CloudSearch.DescribeDomainEndpointOptions

Description

Returns the domain's endpoint options, specifically whether all requests to the domain must arrive over HTTPS. For more information, see Configuring Domain Endpoint Options in the Amazon CloudSearch Developer Guide.

Creating a Request

data DescribeDomainEndpointOptions

Container for the parameters to the DescribeDomainEndpointOptions operation. Specify the name of the domain you want to describe. To show the active configuration and exclude any pending changes, set the Deployed option to true.

See: newDescribeDomainEndpointOptions smart constructor.

Constructors

DescribeDomainEndpointOptions'

Fields

deployed :: Maybe Bool

Whether to retrieve the latest configuration (which might be in a Processing state) or the current, active configuration. Defaults to false.

domainName :: Text

A string that represents the name of a domain.

newDescribeDomainEndpointOptions

Arguments

:: Text
$sel:domainName:DescribeDomainEndpointOptions'

-> DescribeDomainEndpointOptions

Create a value of DescribeDomainEndpointOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deployed:DescribeDomainEndpointOptions', describeDomainEndpointOptions_deployed - Whether to retrieve the latest configuration (which might be in a Processing state) or the current, active configuration. Defaults to false.

$sel:domainName:DescribeDomainEndpointOptions', describeDomainEndpointOptions_domainName - A string that represents the name of a domain.

Request Lenses

describeDomainEndpointOptions_deployed :: Lens' DescribeDomainEndpointOptions (Maybe Bool)

Whether to retrieve the latest configuration (which might be in a Processing state) or the current, active configuration. Defaults to false.

describeDomainEndpointOptions_domainName :: Lens' DescribeDomainEndpointOptions Text

A string that represents the name of a domain.

Destructuring the Response

data DescribeDomainEndpointOptionsResponse

The result of a DescribeDomainEndpointOptions request. Contains the status and configuration of a search domain's endpoint options.

See: newDescribeDomainEndpointOptionsResponse smart constructor.

Constructors

DescribeDomainEndpointOptionsResponse'

Fields

domainEndpointOptions :: Maybe DomainEndpointOptionsStatus

The status and configuration of a search domain's endpoint options.

httpStatus :: Int

The response's http status code.

newDescribeDomainEndpointOptionsResponse

Arguments

:: Int
$sel:httpStatus:DescribeDomainEndpointOptionsResponse'

-> DescribeDomainEndpointOptionsResponse

Create a value of DescribeDomainEndpointOptionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainEndpointOptions:DescribeDomainEndpointOptionsResponse', describeDomainEndpointOptionsResponse_domainEndpointOptions - The status and configuration of a search domain's endpoint options.

$sel:httpStatus:DescribeDomainEndpointOptionsResponse', describeDomainEndpointOptionsResponse_httpStatus - The response's http status code.

Response Lenses

describeDomainEndpointOptionsResponse_domainEndpointOptions :: Lens' DescribeDomainEndpointOptionsResponse (Maybe DomainEndpointOptionsStatus)

The status and configuration of a search domain's endpoint options.

describeDomainEndpointOptionsResponse_httpStatus :: Lens' DescribeDomainEndpointOptionsResponse Int

The response's http status code.