Network.AWS.CloudSearch.DescribeServiceAccessPolicies

Description

Gets information about the access policies that control access to the domain's document and search endpoints. By default, shows the configuration with any pending changes. Set the Deployed option to true to show the active configuration and exclude pending changes. For more information, see Configuring Access for a Search Domain in the Amazon CloudSearch Developer Guide.

Creating a Request

data DescribeServiceAccessPolicies

Container for the parameters to the DescribeServiceAccessPolicies operation. Specifies the name of the domain you want to describe. To show the active configuration and exclude any pending changes, set the Deployed option to true.

See: newDescribeServiceAccessPolicies smart constructor.

Constructors

DescribeServiceAccessPolicies'

Fields

deployed :: Maybe Bool

Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

domainName :: Text

The name of the domain you want to describe.

newDescribeServiceAccessPolicies

Arguments

:: Text
$sel:domainName:DescribeServiceAccessPolicies'

-> DescribeServiceAccessPolicies

Create a value of DescribeServiceAccessPolicies with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deployed:DescribeServiceAccessPolicies', describeServiceAccessPolicies_deployed - Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

$sel:domainName:DescribeServiceAccessPolicies', describeServiceAccessPolicies_domainName - The name of the domain you want to describe.

Request Lenses

describeServiceAccessPolicies_deployed :: Lens' DescribeServiceAccessPolicies (Maybe Bool)

Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

describeServiceAccessPolicies_domainName :: Lens' DescribeServiceAccessPolicies Text

The name of the domain you want to describe.

Destructuring the Response

data DescribeServiceAccessPoliciesResponse

The result of a DescribeServiceAccessPolicies request.

See: newDescribeServiceAccessPoliciesResponse smart constructor.

Constructors

DescribeServiceAccessPoliciesResponse'

Fields

httpStatus :: Int

The response's http status code.

accessPolicies :: AccessPoliciesStatus

The access rules configured for the domain specified in the request.

newDescribeServiceAccessPoliciesResponse

Arguments

:: Int
$sel:httpStatus:DescribeServiceAccessPoliciesResponse'

-> AccessPoliciesStatus
$sel:accessPolicies:DescribeServiceAccessPoliciesResponse'

-> DescribeServiceAccessPoliciesResponse

Create a value of DescribeServiceAccessPoliciesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DescribeServiceAccessPoliciesResponse', describeServiceAccessPoliciesResponse_httpStatus - The response's http status code.

$sel:accessPolicies:DescribeServiceAccessPoliciesResponse', describeServiceAccessPoliciesResponse_accessPolicies - The access rules configured for the domain specified in the request.

Response Lenses

describeServiceAccessPoliciesResponse_httpStatus :: Lens' DescribeServiceAccessPoliciesResponse Int

The response's http status code.

describeServiceAccessPoliciesResponse_accessPolicies :: Lens' DescribeServiceAccessPoliciesResponse AccessPoliciesStatus

The access rules configured for the domain specified in the request.