Network.AWS.CloudSearch.Types.DomainStatus

Description

Documentation

data DomainStatus

The current status of the search domain.

See: newDomainStatus smart constructor.

Constructors

DomainStatus'

Fields

searchInstanceType :: Maybe Text

The instance type that is being used to process search requests.

arn :: Maybe Text

searchPartitionCount :: Maybe Natural

The number of partitions across which the search index is spread.

searchInstanceCount :: Maybe Natural

The number of search instances that are available to process search requests.

limits :: Maybe Limits

processing :: Maybe Bool

True if processing is being done to activate the current domain configuration.

searchService :: Maybe ServiceEndpoint

The service endpoint for requesting search results from a search domain.

created :: Maybe Bool

True if the search domain is created. It can take several minutes to initialize a domain when CreateDomain is called. Newly created search domains are returned from DescribeDomains with a false value for Created until domain creation is complete.

docService :: Maybe ServiceEndpoint

The service endpoint for updating documents in a search domain.

deleted :: Maybe Bool

True if the search domain has been deleted. The system must clean up resources dedicated to the search domain when DeleteDomain is called. Newly deleted search domains are returned from DescribeDomains with a true value for IsDeleted for several minutes until resource cleanup is complete.

domainId :: Text

domainName :: Text

requiresIndexDocuments :: Bool

True if IndexDocuments needs to be called to activate the current domain configuration.

newDomainStatus

Arguments

:: Text
$sel:domainId:DomainStatus'

-> Text
$sel:domainName:DomainStatus'

-> Bool
$sel:requiresIndexDocuments:DomainStatus'

-> DomainStatus

Create a value of DomainStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:searchInstanceType:DomainStatus', domainStatus_searchInstanceType - The instance type that is being used to process search requests.

$sel:arn:DomainStatus', domainStatus_arn - Undocumented member.

$sel:searchPartitionCount:DomainStatus', domainStatus_searchPartitionCount - The number of partitions across which the search index is spread.

$sel:searchInstanceCount:DomainStatus', domainStatus_searchInstanceCount - The number of search instances that are available to process search requests.

$sel:limits:DomainStatus', domainStatus_limits - Undocumented member.

$sel:processing:DomainStatus', domainStatus_processing - True if processing is being done to activate the current domain configuration.

$sel:searchService:DomainStatus', domainStatus_searchService - The service endpoint for requesting search results from a search domain.

$sel:created:DomainStatus', domainStatus_created - True if the search domain is created. It can take several minutes to initialize a domain when CreateDomain is called. Newly created search domains are returned from DescribeDomains with a false value for Created until domain creation is complete.

$sel:docService:DomainStatus', domainStatus_docService - The service endpoint for updating documents in a search domain.

$sel:deleted:DomainStatus', domainStatus_deleted - True if the search domain has been deleted. The system must clean up resources dedicated to the search domain when DeleteDomain is called. Newly deleted search domains are returned from DescribeDomains with a true value for IsDeleted for several minutes until resource cleanup is complete.

$sel:domainId:DomainStatus', domainStatus_domainId - Undocumented member.

$sel:domainName:DomainStatus', domainStatus_domainName - Undocumented member.

$sel:requiresIndexDocuments:DomainStatus', domainStatus_requiresIndexDocuments - True if IndexDocuments needs to be called to activate the current domain configuration.

domainStatus_searchInstanceType :: Lens' DomainStatus (Maybe Text)

The instance type that is being used to process search requests.

domainStatus_arn :: Lens' DomainStatus (Maybe Text)

Undocumented member.

domainStatus_searchPartitionCount :: Lens' DomainStatus (Maybe Natural)

The number of partitions across which the search index is spread.

domainStatus_searchInstanceCount :: Lens' DomainStatus (Maybe Natural)

The number of search instances that are available to process search requests.

domainStatus_limits :: Lens' DomainStatus (Maybe Limits)

Undocumented member.

domainStatus_processing :: Lens' DomainStatus (Maybe Bool)

True if processing is being done to activate the current domain configuration.

domainStatus_searchService :: Lens' DomainStatus (Maybe ServiceEndpoint)

The service endpoint for requesting search results from a search domain.

domainStatus_created :: Lens' DomainStatus (Maybe Bool)

True if the search domain is created. It can take several minutes to initialize a domain when CreateDomain is called. Newly created search domains are returned from DescribeDomains with a false value for Created until domain creation is complete.

domainStatus_docService :: Lens' DomainStatus (Maybe ServiceEndpoint)

The service endpoint for updating documents in a search domain.

domainStatus_deleted :: Lens' DomainStatus (Maybe Bool)

True if the search domain has been deleted. The system must clean up resources dedicated to the search domain when DeleteDomain is called. Newly deleted search domains are returned from DescribeDomains with a true value for IsDeleted for several minutes until resource cleanup is complete.

domainStatus_domainId :: Lens' DomainStatus Text

Undocumented member.

domainStatus_domainName :: Lens' DomainStatus Text

Undocumented member.

domainStatus_requiresIndexDocuments :: Lens' DomainStatus Bool

True if IndexDocuments needs to be called to activate the current domain configuration.