Network.AWS.CloudSearch.Types.IndexField

Description

Documentation

data IndexField

Configuration information for a field in the index, including its name, type, and options. The supported options depend on the IndexFieldType.

See: newIndexField smart constructor.

Constructors

IndexField'

Fields

textArrayOptions :: Maybe TextArrayOptions

latLonOptions :: Maybe LatLonOptions

doubleArrayOptions :: Maybe DoubleArrayOptions

doubleOptions :: Maybe DoubleOptions

dateArrayOptions :: Maybe DateArrayOptions

textOptions :: Maybe TextOptions

literalArrayOptions :: Maybe LiteralArrayOptions

intArrayOptions :: Maybe IntArrayOptions

dateOptions :: Maybe DateOptions

intOptions :: Maybe IntOptions

literalOptions :: Maybe LiteralOptions

indexFieldName :: Text

A string that represents the name of an index field. CloudSearch supports regular index fields as well as dynamic fields. A dynamic field's name defines a pattern that begins or ends with a wildcard. Any document fields that don't map to a regular index field but do match a dynamic field's pattern are configured with the dynamic field's indexing options.

Regular field names begin with a letter and can contain the following characters: a-z (lowercase), 0-9, and _ (underscore). Dynamic field names must begin or end with a wildcard (*). The wildcard can also be the only character in a dynamic field name. Multiple wildcards, and wildcards embedded within a string are not supported.

The name score is reserved and cannot be used as a field name. To reference a document's ID, you can use the name _id.

indexFieldType :: IndexFieldType

newIndexField

Arguments

:: Text
$sel:indexFieldName:IndexField'

-> IndexFieldType
$sel:indexFieldType:IndexField'

-> IndexField

Create a value of IndexField with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:textArrayOptions:IndexField', indexField_textArrayOptions - Undocumented member.

$sel:latLonOptions:IndexField', indexField_latLonOptions - Undocumented member.

$sel:doubleArrayOptions:IndexField', indexField_doubleArrayOptions - Undocumented member.

$sel:doubleOptions:IndexField', indexField_doubleOptions - Undocumented member.

$sel:dateArrayOptions:IndexField', indexField_dateArrayOptions - Undocumented member.

$sel:textOptions:IndexField', indexField_textOptions - Undocumented member.

$sel:literalArrayOptions:IndexField', indexField_literalArrayOptions - Undocumented member.

$sel:intArrayOptions:IndexField', indexField_intArrayOptions - Undocumented member.

$sel:dateOptions:IndexField', indexField_dateOptions - Undocumented member.

$sel:intOptions:IndexField', indexField_intOptions - Undocumented member.

$sel:literalOptions:IndexField', indexField_literalOptions - Undocumented member.

$sel:indexFieldName:IndexField', indexField_indexFieldName - A string that represents the name of an index field. CloudSearch supports regular index fields as well as dynamic fields. A dynamic field's name defines a pattern that begins or ends with a wildcard. Any document fields that don't map to a regular index field but do match a dynamic field's pattern are configured with the dynamic field's indexing options.

Regular field names begin with a letter and can contain the following characters: a-z (lowercase), 0-9, and _ (underscore). Dynamic field names must begin or end with a wildcard (*). The wildcard can also be the only character in a dynamic field name. Multiple wildcards, and wildcards embedded within a string are not supported.

The name score is reserved and cannot be used as a field name. To reference a document's ID, you can use the name _id.

$sel:indexFieldType:IndexField', indexField_indexFieldType - Undocumented member.

indexField_textArrayOptions :: Lens' IndexField (Maybe TextArrayOptions)

Undocumented member.

indexField_latLonOptions :: Lens' IndexField (Maybe LatLonOptions)

Undocumented member.

indexField_doubleArrayOptions :: Lens' IndexField (Maybe DoubleArrayOptions)

Undocumented member.

indexField_doubleOptions :: Lens' IndexField (Maybe DoubleOptions)

Undocumented member.

indexField_dateArrayOptions :: Lens' IndexField (Maybe DateArrayOptions)

Undocumented member.

indexField_textOptions :: Lens' IndexField (Maybe TextOptions)

Undocumented member.

indexField_literalArrayOptions :: Lens' IndexField (Maybe LiteralArrayOptions)

Undocumented member.

indexField_intArrayOptions :: Lens' IndexField (Maybe IntArrayOptions)

Undocumented member.

indexField_dateOptions :: Lens' IndexField (Maybe DateOptions)

Undocumented member.

indexField_intOptions :: Lens' IndexField (Maybe IntOptions)

Undocumented member.

indexField_literalOptions :: Lens' IndexField (Maybe LiteralOptions)

Undocumented member.

indexField_indexFieldName :: Lens' IndexField Text

A string that represents the name of an index field. CloudSearch supports regular index fields as well as dynamic fields. A dynamic field's name defines a pattern that begins or ends with a wildcard. Any document fields that don't map to a regular index field but do match a dynamic field's pattern are configured with the dynamic field's indexing options.

Regular field names begin with a letter and can contain the following characters: a-z (lowercase), 0-9, and _ (underscore). Dynamic field names must begin or end with a wildcard (*). The wildcard can also be the only character in a dynamic field name. Multiple wildcards, and wildcards embedded within a string are not supported.

The name score is reserved and cannot be used as a field name. To reference a document's ID, you can use the name _id.

indexField_indexFieldType :: Lens' IndexField IndexFieldType

Undocumented member.