Network.AWS.CloudSearch.Types.IntArrayOptions

Description

Documentation

data IntArrayOptions

Options for a field that contains an array of 64-bit signed integers. Present if IndexFieldType specifies the field is of type int-array. All options are enabled by default.

See: newIntArrayOptions smart constructor.

Constructors

IntArrayOptions'

Fields

sourceFields :: Maybe Text

A list of source fields to map to the field.

returnEnabled :: Maybe Bool

Whether the contents of the field can be returned in the search results.

facetEnabled :: Maybe Bool

Whether facet information can be returned for the field.

searchEnabled :: Maybe Bool

Whether the contents of the field are searchable.

defaultValue :: Maybe Integer

A value to use for the field if the field isn't specified for a document.

newIntArrayOptions :: IntArrayOptions

Create a value of IntArrayOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceFields:IntArrayOptions', intArrayOptions_sourceFields - A list of source fields to map to the field.

$sel:returnEnabled:IntArrayOptions', intArrayOptions_returnEnabled - Whether the contents of the field can be returned in the search results.

$sel:facetEnabled:IntArrayOptions', intArrayOptions_facetEnabled - Whether facet information can be returned for the field.

$sel:searchEnabled:IntArrayOptions', intArrayOptions_searchEnabled - Whether the contents of the field are searchable.

$sel:defaultValue:IntArrayOptions', intArrayOptions_defaultValue - A value to use for the field if the field isn't specified for a document.

intArrayOptions_sourceFields :: Lens' IntArrayOptions (Maybe Text)

A list of source fields to map to the field.

intArrayOptions_returnEnabled :: Lens' IntArrayOptions (Maybe Bool)

Whether the contents of the field can be returned in the search results.

intArrayOptions_facetEnabled :: Lens' IntArrayOptions (Maybe Bool)

Whether facet information can be returned for the field.

intArrayOptions_searchEnabled :: Lens' IntArrayOptions (Maybe Bool)

Whether the contents of the field are searchable.

intArrayOptions_defaultValue :: Lens' IntArrayOptions (Maybe Integer)

A value to use for the field if the field isn't specified for a document.