Network.AWS.CloudSearch.Types.LatLonOptions

Description

Documentation

data LatLonOptions

Options for a latlon field. A latlon field contains a location stored as a latitude and longitude value pair. Present if IndexFieldType specifies the field is of type latlon. All options are enabled by default.

See: newLatLonOptions smart constructor.

Constructors

LatLonOptions'

Fields

sortEnabled :: Maybe Bool

Whether the field can be used to sort the search results.

facetEnabled :: Maybe Bool

Whether facet information can be returned for the field.

returnEnabled :: Maybe Bool

Whether the contents of the field can be returned in the search results.

sourceField :: Maybe Text

searchEnabled :: Maybe Bool

Whether the contents of the field are searchable.

defaultValue :: Maybe Text

A value to use for the field if the field isn't specified for a document.

newLatLonOptions :: LatLonOptions

Create a value of LatLonOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sortEnabled:LatLonOptions', latLonOptions_sortEnabled - Whether the field can be used to sort the search results.

$sel:facetEnabled:LatLonOptions', latLonOptions_facetEnabled - Whether facet information can be returned for the field.

$sel:returnEnabled:LatLonOptions', latLonOptions_returnEnabled - Whether the contents of the field can be returned in the search results.

$sel:sourceField:LatLonOptions', latLonOptions_sourceField - Undocumented member.

$sel:searchEnabled:LatLonOptions', latLonOptions_searchEnabled - Whether the contents of the field are searchable.

$sel:defaultValue:LatLonOptions', latLonOptions_defaultValue - A value to use for the field if the field isn't specified for a document.

latLonOptions_sortEnabled :: Lens' LatLonOptions (Maybe Bool)

Whether the field can be used to sort the search results.

latLonOptions_facetEnabled :: Lens' LatLonOptions (Maybe Bool)

Whether facet information can be returned for the field.

latLonOptions_returnEnabled :: Lens' LatLonOptions (Maybe Bool)

Whether the contents of the field can be returned in the search results.

latLonOptions_sourceField :: Lens' LatLonOptions (Maybe Text)

Undocumented member.

latLonOptions_searchEnabled :: Lens' LatLonOptions (Maybe Bool)

Whether the contents of the field are searchable.

latLonOptions_defaultValue :: Lens' LatLonOptions (Maybe Text)

A value to use for the field if the field isn't specified for a document.