Network.AWS.CloudSearch.Types.Limits

Description

Documentation

data Limits

See: newLimits smart constructor.

Constructors

Limits'

Fields

maximumReplicationCount :: Natural

maximumPartitionCount :: Natural

newLimits

Arguments

:: Natural
$sel:maximumReplicationCount:Limits'

-> Natural
$sel:maximumPartitionCount:Limits'

-> Limits

Create a value of Limits with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maximumReplicationCount:Limits', limits_maximumReplicationCount - Undocumented member.

$sel:maximumPartitionCount:Limits', limits_maximumPartitionCount - Undocumented member.

limits_maximumReplicationCount :: Lens' Limits Natural

Undocumented member.

limits_maximumPartitionCount :: Lens' Limits Natural

Undocumented member.