Network.AWS.CloudSearch.Types.Suggester

Description

Documentation

data Suggester

Configuration information for a search suggester. Each suggester has a unique name and specifies the text field you want to use for suggestions. The following options can be configured for a suggester: FuzzyMatching, SortExpression.

See: newSuggester smart constructor.

Constructors

Suggester'

Fields

suggesterName :: Text

documentSuggesterOptions :: DocumentSuggesterOptions

newSuggester

Arguments

:: Text
$sel:suggesterName:Suggester'

-> DocumentSuggesterOptions
$sel:documentSuggesterOptions:Suggester'

-> Suggester

Create a value of Suggester with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:suggesterName:Suggester', suggester_suggesterName - Undocumented member.

$sel:documentSuggesterOptions:Suggester', suggester_documentSuggesterOptions - Undocumented member.

suggester_suggesterName :: Lens' Suggester Text

Undocumented member.

suggester_documentSuggesterOptions :: Lens' Suggester DocumentSuggesterOptions

Undocumented member.