Network.AWS.CloudSearch.Types.TextOptions

Description

Documentation

data TextOptions

Options for text field. Present if IndexFieldType specifies the field is of type text. A text field is always searchable. All options are enabled by default.

See: newTextOptions smart constructor.

Constructors

TextOptions'

Fields

sortEnabled :: Maybe Bool

Whether the field can be used to sort the search results.

analysisScheme :: Maybe Text

The name of an analysis scheme for a text field.

returnEnabled :: Maybe Bool

Whether the contents of the field can be returned in the search results.

sourceField :: Maybe Text

defaultValue :: Maybe Text

A value to use for the field if the field isn't specified for a document.

highlightEnabled :: Maybe Bool

Whether highlights can be returned for the field.

newTextOptions :: TextOptions

Create a value of TextOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sortEnabled:TextOptions', textOptions_sortEnabled - Whether the field can be used to sort the search results.

$sel:analysisScheme:TextOptions', textOptions_analysisScheme - The name of an analysis scheme for a text field.

$sel:returnEnabled:TextOptions', textOptions_returnEnabled - Whether the contents of the field can be returned in the search results.

$sel:sourceField:TextOptions', textOptions_sourceField - Undocumented member.

$sel:defaultValue:TextOptions', textOptions_defaultValue - A value to use for the field if the field isn't specified for a document.

$sel:highlightEnabled:TextOptions', textOptions_highlightEnabled - Whether highlights can be returned for the field.

textOptions_sortEnabled :: Lens' TextOptions (Maybe Bool)

Whether the field can be used to sort the search results.

textOptions_analysisScheme :: Lens' TextOptions (Maybe Text)

The name of an analysis scheme for a text field.

textOptions_returnEnabled :: Lens' TextOptions (Maybe Bool)

Whether the contents of the field can be returned in the search results.

textOptions_sourceField :: Lens' TextOptions (Maybe Text)

Undocumented member.

textOptions_defaultValue :: Lens' TextOptions (Maybe Text)

A value to use for the field if the field isn't specified for a document.

textOptions_highlightEnabled :: Lens' TextOptions (Maybe Bool)

Whether highlights can be returned for the field.