Network.AWS.CloudSearch.UpdateDomainEndpointOptions

Description

Updates the domain's endpoint options, specifically whether all requests to the domain must arrive over HTTPS. For more information, see Configuring Domain Endpoint Options in the Amazon CloudSearch Developer Guide.

Creating a Request

data UpdateDomainEndpointOptions

Container for the parameters to the UpdateDomainEndpointOptions operation. Specifies the name of the domain you want to update and the domain endpoint options.

See: newUpdateDomainEndpointOptions smart constructor.

Constructors

UpdateDomainEndpointOptions'

Fields

domainName :: Text

A string that represents the name of a domain.

domainEndpointOptions :: DomainEndpointOptions

Whether to require that all requests to the domain arrive over HTTPS. We recommend Policy-Min-TLS-1-2-2019-07 for TLSSecurityPolicy. For compatibility with older clients, the default is Policy-Min-TLS-1-0-2019-07.

newUpdateDomainEndpointOptions

Arguments

:: Text
$sel:domainName:UpdateDomainEndpointOptions'

-> DomainEndpointOptions
$sel:domainEndpointOptions:UpdateDomainEndpointOptions'

-> UpdateDomainEndpointOptions

Create a value of UpdateDomainEndpointOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainName:UpdateDomainEndpointOptions', updateDomainEndpointOptions_domainName - A string that represents the name of a domain.

$sel:domainEndpointOptions:UpdateDomainEndpointOptions', updateDomainEndpointOptions_domainEndpointOptions - Whether to require that all requests to the domain arrive over HTTPS. We recommend Policy-Min-TLS-1-2-2019-07 for TLSSecurityPolicy. For compatibility with older clients, the default is Policy-Min-TLS-1-0-2019-07.

Request Lenses

updateDomainEndpointOptions_domainName :: Lens' UpdateDomainEndpointOptions Text

A string that represents the name of a domain.

updateDomainEndpointOptions_domainEndpointOptions :: Lens' UpdateDomainEndpointOptions DomainEndpointOptions

Whether to require that all requests to the domain arrive over HTTPS. We recommend Policy-Min-TLS-1-2-2019-07 for TLSSecurityPolicy. For compatibility with older clients, the default is Policy-Min-TLS-1-0-2019-07.

Destructuring the Response

data UpdateDomainEndpointOptionsResponse

The result of a UpdateDomainEndpointOptions request. Contains the configuration and status of the domain's endpoint options.

See: newUpdateDomainEndpointOptionsResponse smart constructor.

Constructors

UpdateDomainEndpointOptionsResponse'

Fields

domainEndpointOptions :: Maybe DomainEndpointOptionsStatus

The newly-configured domain endpoint options.

httpStatus :: Int

The response's http status code.

newUpdateDomainEndpointOptionsResponse

Arguments

:: Int
$sel:httpStatus:UpdateDomainEndpointOptionsResponse'

-> UpdateDomainEndpointOptionsResponse

Create a value of UpdateDomainEndpointOptionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainEndpointOptions:UpdateDomainEndpointOptions', updateDomainEndpointOptionsResponse_domainEndpointOptions - The newly-configured domain endpoint options.

$sel:httpStatus:UpdateDomainEndpointOptionsResponse', updateDomainEndpointOptionsResponse_httpStatus - The response's http status code.

Response Lenses

updateDomainEndpointOptionsResponse_domainEndpointOptions :: Lens' UpdateDomainEndpointOptionsResponse (Maybe DomainEndpointOptionsStatus)

The newly-configured domain endpoint options.

updateDomainEndpointOptionsResponse_httpStatus :: Lens' UpdateDomainEndpointOptionsResponse Int

The response's http status code.