Network.AWS.CloudSearch.UpdateScalingParameters

Description

Configures scaling parameters for a domain. A domain's scaling parameters specify the desired search instance type and replication count. Amazon CloudSearch will still automatically scale your domain based on the volume of data and traffic, but not below the desired instance type and replication count. If the Multi-AZ option is enabled, these values control the resources used per Availability Zone. For more information, see Configuring Scaling Options in the Amazon CloudSearch Developer Guide.

Creating a Request

data UpdateScalingParameters

Container for the parameters to the UpdateScalingParameters operation. Specifies the name of the domain you want to update and the scaling parameters you want to configure.

See: newUpdateScalingParameters smart constructor.

Constructors

UpdateScalingParameters'

Fields

domainName :: Text

scalingParameters :: ScalingParameters

newUpdateScalingParameters

Arguments

:: Text
$sel:domainName:UpdateScalingParameters'

-> ScalingParameters
$sel:scalingParameters:UpdateScalingParameters'

-> UpdateScalingParameters

Create a value of UpdateScalingParameters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainName:UpdateScalingParameters', updateScalingParameters_domainName - Undocumented member.

$sel:scalingParameters:UpdateScalingParameters', updateScalingParameters_scalingParameters - Undocumented member.

Request Lenses

updateScalingParameters_domainName :: Lens' UpdateScalingParameters Text

Undocumented member.

updateScalingParameters_scalingParameters :: Lens' UpdateScalingParameters ScalingParameters

Undocumented member.

Destructuring the Response

data UpdateScalingParametersResponse

The result of a UpdateScalingParameters request. Contains the status of the newly-configured scaling parameters.

See: newUpdateScalingParametersResponse smart constructor.

Constructors

UpdateScalingParametersResponse'

Fields

httpStatus :: Int

The response's http status code.

scalingParameters :: ScalingParametersStatus

newUpdateScalingParametersResponse

Arguments

:: Int
$sel:httpStatus:UpdateScalingParametersResponse'

-> ScalingParametersStatus
$sel:scalingParameters:UpdateScalingParameters'

-> UpdateScalingParametersResponse

Create a value of UpdateScalingParametersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateScalingParametersResponse', updateScalingParametersResponse_httpStatus - The response's http status code.

$sel:scalingParameters:UpdateScalingParameters', updateScalingParametersResponse_scalingParameters - Undocumented member.

Response Lenses

updateScalingParametersResponse_httpStatus :: Lens' UpdateScalingParametersResponse Int

The response's http status code.

updateScalingParametersResponse_scalingParameters :: Lens' UpdateScalingParametersResponse ScalingParametersStatus

Undocumented member.