Network.AWS.CloudSearch.V2013_01_01.CreateDomain

Description

Creates a new search domain. For more information, see Creating a Search Domain in the Amazon CloudSearch Developer Guide.

Documentation

data CreateDomain

Constructors

CreateDomain

Fields

_cdrDomainName :: Text

A name for the domain you are creating. Allowed characters are a-z (lower-case letters), 0-9, and hyphen (-). Domain names must start with a letter or number and be at least 3 and no more than 28 characters long.

cdrDomainName :: Iso' CreateDomain Text

data CreateDomainResponse

Constructors

CreateDomainResponse

Fields

_cdsDomainStatus :: Maybe DomainStatus

The current status of the search domain.

cdsDomainStatus :: Iso' CreateDomainResponse (Maybe DomainStatus)