Network.AWS.CloudSearch.V2013_01_01.DefineAnalysisScheme

Description

Configures an analysis scheme that can be applied to a text or text-array field to define language-specific text processing options. For more information, see Configuring Analysis Schemes in the Amazon CloudSearch Developer Guide.

Documentation

data DefineAnalysisScheme

Constructors

DefineAnalysisScheme

Fields

_dasvAnalysisScheme :: AnalysisScheme

Configuration information for an analysis scheme. Each analysis scheme has a unique name and specifies the language of the text to be processed. The following options can be configured for an analysis scheme: Synonyms, Stopwords, StemmingDictionary, and AlgorithmicStemming.

_dasvDomainName :: Text

A string that represents the name of a domain. Domain names are unique across the domains owned by an account within an AWS region. Domain names start with a letter or number and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

dasvDomainName :: Lens' DefineAnalysisScheme Text

dasvAnalysisScheme :: Lens' DefineAnalysisScheme AnalysisScheme

data DefineAnalysisSchemeResponse

Constructors

DefineAnalysisSchemeResponse

Fields

_daswAnalysisScheme :: AnalysisSchemeStatus

The status and configuration of an AnalysisScheme.

daswAnalysisScheme :: Iso' DefineAnalysisSchemeResponse AnalysisSchemeStatus