Network.AWS.CloudSearch.V2013_01_01.DefineSuggester

Description

Configures a suggester for a domain. A suggester enables you to display possible matches before users finish typing their queries. When you configure a suggester, you must specify the name of the text field you want to search for possible matches and a unique name for the suggester. For more information, see Getting Search Suggestions in the Amazon CloudSearch Developer Guide.

Documentation

data DefineSuggester

Constructors

DefineSuggester

Fields

_dstDomainName :: Text

A string that represents the name of a domain. Domain names are unique across the domains owned by an account within an AWS region. Domain names start with a letter or number and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

_dstSuggester :: Suggester

Configuration information for a search suggester. Each suggester has a unique name and specifies the text field you want to use for suggestions. The following options can be configured for a suggester: FuzzyMatching, SortExpression.

dstSuggester :: Lens' DefineSuggester Suggester

dstDomainName :: Lens' DefineSuggester Text

data DefineSuggesterResponse

Constructors

DefineSuggesterResponse

Fields

_dsuSuggester :: SuggesterStatus

The value of a Suggester and its current status.

dsuSuggester :: Iso' DefineSuggesterResponse SuggesterStatus