Network.AWS.CloudSearch.V2013_01_01.DeleteExpression

Description

Removes an Expression from the search domain. For more information, see Configuring Expressions in the Amazon CloudSearch Developer Guide.

Documentation

data DeleteExpression

Constructors

DeleteExpression

Fields

_devDomainName :: Text

A string that represents the name of a domain. Domain names are unique across the domains owned by an account within an AWS region. Domain names start with a letter or number and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

_devExpressionName :: Text

The name of the Expression to delete.

devExpressionName :: Lens' DeleteExpression Text

devDomainName :: Lens' DeleteExpression Text

data DeleteExpressionResponse

Constructors

DeleteExpressionResponse

Fields

_dewExpression :: ExpressionStatus

The status of the expression being deleted.

dewExpression :: Iso' DeleteExpressionResponse ExpressionStatus