Network.AWS.CloudSearch.V2013_01_01.DeleteSuggester

Description

Deletes a suggester. For more information, see Getting Search Suggestions in the Amazon CloudSearch Developer Guide.

Documentation

data DeleteSuggester

Constructors

DeleteSuggester

Fields

_dsvDomainName :: Text

A string that represents the name of a domain. Domain names are unique across the domains owned by an account within an AWS region. Domain names start with a letter or number and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

_dsvSuggesterName :: Text

Specifies the name of the suggester you want to delete.

dsvSuggesterName :: Lens' DeleteSuggester Text

dsvDomainName :: Lens' DeleteSuggester Text

data DeleteSuggesterResponse

Constructors

DeleteSuggesterResponse

Fields

_dswSuggester :: SuggesterStatus

The status of the suggester being deleted.

dswSuggester :: Iso' DeleteSuggesterResponse SuggesterStatus