Network.AWS.CloudSearch.V2013_01_01.DescribeAnalysisSchemes

Description

Gets the analysis schemes configured for a domain. An analysis scheme defines language-specific text processing options for a text field. Can be limited to specific analysis schemes by name. By default, shows all analysis schemes and includes any pending changes to the configuration. Set the Deployed option to true to show the active configuration and exclude pending changes. For more information, see Configuring Analysis Schemes in the Amazon CloudSearch Developer Guide.

Documentation

describeAnalysisSchemes

Arguments

:: Text
_dastDomainName

-> DescribeAnalysisSchemes

Minimum specification for a DescribeAnalysisSchemes request.

data DescribeAnalysisSchemes

Constructors

DescribeAnalysisSchemes

Fields

_dastDomainName :: Text

The name of the domain you want to describe.

_dastDeployed :: Maybe Bool

Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

_dastAnalysisSchemeNames :: [Text]

The analysis schemes you want to describe.

dastDomainName :: Lens' DescribeAnalysisSchemes Text

dastDeployed :: Lens' DescribeAnalysisSchemes (Maybe Bool)

dastAnalysisSchemeNames :: Lens' DescribeAnalysisSchemes [Text]

data DescribeAnalysisSchemesResponse

Constructors

DescribeAnalysisSchemesResponse

Fields

_dasuAnalysisSchemes :: [AnalysisSchemeStatus]

The analysis scheme descriptions.

dasuAnalysisSchemes :: Iso' DescribeAnalysisSchemesResponse [AnalysisSchemeStatus]