Network.AWS.CloudSearch.V2013_01_01.DescribeAvailabilityOptions

Description

Gets the availability options configured for a domain. By default, shows the configuration with any pending changes. Set the Deployed option to true to show the active configuration and exclude pending changes. For more information, see Configuring Availability Options in the Amazon CloudSearch Developer Guide.

Documentation

describeAvailabilityOptions

Arguments

:: Text
_daorDomainName

-> DescribeAvailabilityOptions

Minimum specification for a DescribeAvailabilityOptions request.

data DescribeAvailabilityOptions

Constructors

DescribeAvailabilityOptions

Fields

_daorDomainName :: Text

The name of the domain you want to describe.

_daorDeployed :: Maybe Bool

Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

daorDomainName :: Lens' DescribeAvailabilityOptions Text

daorDeployed :: Lens' DescribeAvailabilityOptions (Maybe Bool)

data DescribeAvailabilityOptionsResponse

Constructors

DescribeAvailabilityOptionsResponse

Fields

_daosAvailabilityOptions :: Maybe AvailabilityOptionsStatus

The availability options configured for the domain. Indicates whether Multi-AZ is enabled for the domain.

daosAvailabilityOptions :: Iso' DescribeAvailabilityOptionsResponse (Maybe AvailabilityOptionsStatus)