Network.AWS.CloudSearch.V2013_01_01.DescribeExpressions

Description

Gets the expressions configured for the search domain. Can be limited to specific expressions by name. By default, shows all expressions and includes any pending changes to the configuration. Set the Deployed option to true to show the active configuration and exclude pending changes. For more information, see Configuring Expressions in the Amazon CloudSearch Developer Guide.

Documentation

describeExpressions

Arguments

:: Text
_derDomainName

-> DescribeExpressions

Minimum specification for a DescribeExpressions request.

data DescribeExpressions

Constructors

DescribeExpressions

Fields

_derDomainName :: Text

The name of the domain you want to describe.

_derDeployed :: Maybe Bool

Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

_derExpressionNames :: [Text]

Limits the DescribeExpressions response to the specified expressions. If not specified, all expressions are shown.

derExpressionNames :: Lens' DescribeExpressions [Text]

derDomainName :: Lens' DescribeExpressions Text

derDeployed :: Lens' DescribeExpressions (Maybe Bool)

data DescribeExpressionsResponse

Constructors

DescribeExpressionsResponse

Fields

_desExpressions :: [ExpressionStatus]

The expressions configured for the domain.

desExpressions :: Iso' DescribeExpressionsResponse [ExpressionStatus]