Network.AWS.CloudSearch.V2013_01_01.DescribeScalingParameters

Description

Gets the scaling parameters configured for a domain. A domain's scaling parameters specify the desired search instance type and replication count. For more information, see Configuring Scaling Options in the Amazon CloudSearch Developer Guide.

Documentation

data DescribeScalingParameters

Constructors

DescribeScalingParameters

Fields

_dsprDomainName :: Text

A string that represents the name of a domain. Domain names are unique across the domains owned by an account within an AWS region. Domain names start with a letter or number and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

dsprDomainName :: Iso' DescribeScalingParameters Text

data DescribeScalingParametersResponse

Constructors

DescribeScalingParametersResponse

Fields

_dspsScalingParameters :: ScalingParametersStatus

The status and configuration of a search domain's scaling parameters.

dspsScalingParameters :: Iso' DescribeScalingParametersResponse ScalingParametersStatus