Network.AWS.CloudSearch.V2013_01_01.DescribeSuggesters

Description

Gets the suggesters configured for a domain. A suggester enables you to display possible matches before users finish typing their queries. Can be limited to specific suggesters by name. By default, shows all suggesters and includes any pending changes to the configuration. Set the Deployed option to true to show the active configuration and exclude pending changes. For more information, see Getting Search Suggestions in the Amazon CloudSearch Developer Guide.

Documentation

describeSuggesters

Arguments

:: Text
_dsrDomainName

-> DescribeSuggesters

Minimum specification for a DescribeSuggesters request.

data DescribeSuggesters

Constructors

DescribeSuggesters

Fields

_dsrDomainName :: Text

The name of the domain you want to describe.

_dsrDeployed :: Maybe Bool

Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

_dsrSuggesterNames :: [Text]

The suggesters you want to describe.

dsrSuggesterNames :: Lens' DescribeSuggesters [Text]

dsrDomainName :: Lens' DescribeSuggesters Text

dsrDeployed :: Lens' DescribeSuggesters (Maybe Bool)

data DescribeSuggestersResponse

Constructors

DescribeSuggestersResponse

Fields

_dssSuggesters :: [SuggesterStatus]

The suggesters configured for the domain specified in the request.

dssSuggesters :: Iso' DescribeSuggestersResponse [SuggesterStatus]