Network.AWS.CloudSearch.V2013_01_01.IndexDocuments

Description

Tells the search domain to start indexing its documents using the latest indexing options. This operation must be invoked to activate options whose OptionStatus is RequiresIndexDocuments.

Documentation

data IndexDocuments

Constructors

IndexDocuments

Fields

_idrDomainName :: Text

A string that represents the name of a domain. Domain names are unique across the domains owned by an account within an AWS region. Domain names start with a letter or number and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

idrDomainName :: Iso' IndexDocuments Text

data IndexDocumentsResponse

Constructors

IndexDocumentsResponse

Fields

_idsFieldNames :: [Text]

The names of the fields that are currently being indexed.

idsFieldNames :: Iso' IndexDocumentsResponse [Text]