Network.AWS.CloudSearchDomains

Description

Derived from API version 2013-01-01 of the AWS service descriptions, licensed under Apache 2.0.

You use the AmazonCloudSearch2013 API to upload documents to a search domain and search those documents.

The endpoints for submitting UploadDocuments, Search, and Suggest requests are domain-specific. To get the endpoints for your domain, use the Amazon CloudSearch configuration service DescribeDomains action. The domain endpoints are also displayed on the domain dashboard in the Amazon CloudSearch console. You submit suggest requests to the search endpoint.

For more information, see the Amazon CloudSearch Developer Guide.

Service Configuration

defaultService :: Service

API version 2013-01-01 of the Amazon CloudSearch Domain SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by CloudSearchDomains.

SearchException

_SearchException :: AsError a => Getting (First ServiceError) a ServiceError

Information about any problems encountered while processing a search request.

DocumentServiceException

_DocumentServiceException :: AsError a => Getting (First ServiceError) a ServiceError

Information about any problems encountered while processing an upload request.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

UploadDocuments

data UploadDocuments

Container for the parameters to the UploadDocuments request.

See: newUploadDocuments smart constructor.

Constructors

UploadDocuments' ContentType HashedBody

newUploadDocuments

Arguments

:: ContentType
$sel:contentType:UploadDocuments'

-> HashedBody
$sel:documents:UploadDocuments'

-> UploadDocuments

Create a value of UploadDocuments with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:contentType:UploadDocuments', uploadDocuments_contentType - The format of the batch you are uploading. Amazon CloudSearch supports two document batch formats:

application/json

application/xml

$sel:documents:UploadDocuments', uploadDocuments_documents - A batch of documents formatted in JSON or HTML.

data UploadDocumentsResponse

Contains the response to an UploadDocuments request.

See: newUploadDocumentsResponse smart constructor.

Constructors

UploadDocumentsResponse' (Maybe Text) (Maybe [DocumentServiceWarning]) (Maybe Integer) (Maybe Integer) Int

newUploadDocumentsResponse

Arguments

:: Int
$sel:httpStatus:UploadDocumentsResponse'

-> UploadDocumentsResponse

Create a value of UploadDocumentsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:UploadDocumentsResponse', uploadDocumentsResponse_status - The status of an UploadDocumentsRequest.

$sel:warnings:UploadDocumentsResponse', uploadDocumentsResponse_warnings - Any warnings returned by the document service about the documents being uploaded.

$sel:deletes:UploadDocumentsResponse', uploadDocumentsResponse_deletes - The number of documents that were deleted from the search domain.

$sel:adds:UploadDocumentsResponse', uploadDocumentsResponse_adds - The number of documents that were added to the search domain.

$sel:httpStatus:UploadDocumentsResponse', uploadDocumentsResponse_httpStatus - The response's http status code.

Search

data Search

Container for the parameters to the Search request.

See: newSearch smart constructor.

Constructors

Search' (Maybe Text) (Maybe Bool) (Maybe QueryParser) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Integer) (Maybe Text) (Maybe Integer) (Maybe Text) (Maybe Text) Text

newSearch

Arguments

:: Text
$sel:query:Search'

-> Search

Create a value of Search with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expr:Search', search_expr - Defines one or more numeric expressions that can be used to sort results or specify search or filter criteria. You can also specify expressions as return fields.

You specify the expressions in JSON using the form {"EXPRESSIONNAME":"EXPRESSION"}. You can define and use multiple expressions in a search request. For example:

 {"expression1":"_score*rating", "expression2":"(1/rank)*year"}

For information about the variables, operators, and functions you can use in expressions, see Writing Expressions in the Amazon CloudSearch Developer Guide.

$sel:partial:Search', search_partial - Enables partial results to be returned if one or more index partitions are unavailable. When your search index is partitioned across multiple search instances, by default Amazon CloudSearch only returns results if every partition can be queried. This means that the failure of a single search instance can result in 5xx (internal server) errors. When you enable partial results, Amazon CloudSearch returns whatever results are available and includes the percentage of documents searched in the search results (percent-searched). This enables you to more gracefully degrade your users' search experience. For example, rather than displaying no results, you could display the partial results and a message indicating that the results might be incomplete due to a temporary system outage.

$sel:queryParser:Search', search_queryParser - Specifies which query parser to use to process the request. If queryParser is not specified, Amazon CloudSearch uses the simple query parser.

Amazon CloudSearch supports four query parsers:

simple: perform simple searches of text and text-array fields. By default, the simple query parser searches all text and text-array fields. You can specify which fields to search by with the queryOptions parameter. If you prefix a search term with a plus sign (+) documents must contain the term to be considered a match. (This is the default, unless you configure the default operator with the queryOptions parameter.) You can use the - (NOT), | (OR), and * (wildcard) operators to exclude particular terms, find results that match any of the specified terms, or search for a prefix. To search for a phrase rather than individual terms, enclose the phrase in double quotes. For more information, see Searching for Text in the Amazon CloudSearch Developer Guide.

structured: perform advanced searches by combining multiple expressions to define the search criteria. You can also search within particular fields, search for values and ranges of values, and use advanced options such as term boosting, matchall, and near. For more information, see Constructing Compound Queries in the Amazon CloudSearch Developer Guide.

lucene: search using the Apache Lucene query parser syntax. For more information, see Apache Lucene Query Parser Syntax.

dismax: search using the simplified subset of the Apache Lucene query parser syntax defined by the DisMax query parser. For more information, see DisMax Query Parser Syntax.

$sel:queryOptions:Search', search_queryOptions - Configures options for the query parser specified in the queryParser parameter. You specify the options in JSON using the following form {"OPTION1":"VALUE1","OPTION2":VALUE2"..."OPTIONN":"VALUEN"}.

The options you can configure vary according to which parser you use:

defaultOperator: The default operator used to combine individual terms in the search string. For example: defaultOperator: 'or'. For the dismax parser, you specify a percentage that represents the percentage of terms in the search string (rounded down) that must match, rather than a default operator. A value of 0% is the equivalent to OR, and a value of 100% is equivalent to AND. The percentage must be specified as a value in the range 0-100 followed by the percent (%) symbol. For example, defaultOperator: 50%. Valid values: and, or, a percentage in the range 0%-100% (dismax). Default: and (simple, structured, lucene) or 100 (dismax). Valid for: simple, structured, lucene, and dismax.

fields: An array of the fields to search when no fields are specified in a search. If no fields are specified in a search and this option is not specified, all text and text-array fields are searched. You can specify a weight for each field to control the relative importance of each field when Amazon CloudSearch calculates relevance scores. To specify a field weight, append a caret (^) symbol and the weight to the field name. For example, to boost the importance of the title field over the description field you could specify: "fields":["title^5","description"]. Valid values: The name of any configured field and an optional numeric value greater than zero. Default: All text and text-array fields. Valid for: simple, structured, lucene, and dismax.

operators: An array of the operators or special characters you want to disable for the simple query parser. If you disable the and, or, or not operators, the corresponding operators (+, |, -) have no special meaning and are dropped from the search string. Similarly, disabling prefix disables the wildcard operator (*) and disabling phrase disables the ability to search for phrases by enclosing phrases in double quotes. Disabling precedence disables the ability to control order of precedence using parentheses. Disabling near disables the ability to use the ~ operator to perform a sloppy phrase search. Disabling the fuzzy operator disables the ability to use the ~ operator to perform a fuzzy search. escape disables the ability to use a backslash (\) to escape special characters within the search string. Disabling whitespace is an advanced option that prevents the parser from tokenizing on whitespace, which can be useful for Vietnamese. (It prevents Vietnamese words from being split incorrectly.) For example, you could disable all operators other than the phrase operator to support just simple term and phrase queries: "operators":["and","not","or", "prefix"]. Valid values: and, escape, fuzzy, near, not, or, phrase, precedence, prefix, whitespace. Default: All operators and special characters are enabled. Valid for: simple.

phraseFields: An array of the text or text-array fields you want to use for phrase searches. When the terms in the search string appear in close proximity within a field, the field scores higher. You can specify a weight for each field to boost that score. The phraseSlop option controls how much the matches can deviate from the search string and still be boosted. To specify a field weight, append a caret (^) symbol and the weight to the field name. For example, to boost phrase matches in the title field over the abstract field, you could specify: "phraseFields":["title^3", "plot"] Valid values: The name of any text or text-array field and an optional numeric value greater than zero. Default: No fields. If you don't specify any fields with phraseFields, proximity scoring is disabled even if phraseSlop is specified. Valid for: dismax.

phraseSlop: An integer value that specifies how much matches can deviate from the search phrase and still be boosted according to the weights specified in the phraseFields option; for example, phraseSlop: 2. You must also specify phraseFields to enable proximity scoring. Valid values: positive integers. Default: 0. Valid for: dismax.

explicitPhraseSlop: An integer value that specifies how much a match can deviate from the search phrase when the phrase is enclosed in double quotes in the search string. (Phrases that exceed this proximity distance are not considered a match.) For example, to specify a slop of three for dismax phrase queries, you would specify "explicitPhraseSlop":3. Valid values: positive integers. Default: 0. Valid for: dismax.

tieBreaker: When a term in the search string is found in a document's field, a score is calculated for that field based on how common the word is in that field compared to other documents. If the term occurs in multiple fields within a document, by default only the highest scoring field contributes to the document's overall score. You can specify a tieBreaker value to enable the matches in lower-scoring fields to contribute to the document's score. That way, if two documents have the same max field score for a particular term, the score for the document that has matches in more fields will be higher. The formula for calculating the score with a tieBreaker is (max field score) + (tieBreaker) * (sum of the scores for the rest of the matching fields). Set tieBreaker to 0 to disregard all but the highest scoring field (pure max): "tieBreaker":0. Set to 1 to sum the scores from all fields (pure sum): "tieBreaker":1. Valid values: 0.0 to 1.0. Default: 0.0. Valid for: dismax.

$sel:return':Search', search_return - Specifies the field and expression values to include in the response. Multiple fields or expressions are specified as a comma-separated list. By default, a search response includes all return enabled fields (_all_fields). To return only the document IDs for the matching documents, specify _no_fields. To retrieve the relevance score calculated for each document, specify _score.

$sel:cursor:Search', search_cursor - Retrieves a cursor value you can use to page through large result sets. Use the size parameter to control the number of hits to include in each response. You can specify either the cursor or start parameter in a request; they are mutually exclusive. To get the first cursor, set the cursor value to initial. In subsequent requests, specify the cursor value returned in the hits section of the response.

For more information, see Paginating Results in the Amazon CloudSearch Developer Guide.

$sel:stats:Search', search_stats - Specifies one or more fields for which to get statistics information. Each specified field must be facet-enabled in the domain configuration. The fields are specified in JSON using the form:

{"FIELD-A":{},"FIELD-B":{}}

There are currently no options supported for statistics.

$sel:highlight:Search', search_highlight - Retrieves highlights for matches in the specified text or text-array fields. Each specified field must be highlight enabled in the domain configuration. The fields and options are specified in JSON using the form {"FIELD":{"OPTION":VALUE,"OPTION:"STRING"},"FIELD":{"OPTION":VALUE,"OPTION":"STRING"}}.

You can specify the following highlight options:

format: specifies the format of the data in the text field: text or html. When data is returned as HTML, all non-alphanumeric characters are encoded. The default is html.

max_phrases: specifies the maximum number of occurrences of the search term(s) you want to highlight. By default, the first occurrence is highlighted.

pre_tag: specifies the string to prepend to an occurrence of a search term. The default for HTML highlights is &lt;em&gt;. The default for text highlights is *.

post_tag: specifies the string to append to an occurrence of a search term. The default for HTML highlights is &lt;/em&gt;. The default for text highlights is *.

If no highlight options are specified for a field, the returned field text is treated as HTML and the first match is highlighted with emphasis tags: &lt;em>search-term&lt;/em&gt;.

For example, the following request retrieves highlights for the actors and title fields.

{ "actors": {}, "title": {"format": "text","max_phrases": 2,"pre_tag": "","post_tag": ""} }

$sel:start:Search', search_start - Specifies the offset of the first search hit you want to return. Note that the result set is zero-based; the first result is at index 0. You can specify either the start or cursor parameter in a request, they are mutually exclusive.

For more information, see Paginating Results in the Amazon CloudSearch Developer Guide.

$sel:facet:Search', search_facet - Specifies one or more fields for which to get facet information, and options that control how the facet information is returned. Each specified field must be facet-enabled in the domain configuration. The fields and options are specified in JSON using the form {"FIELD":{"OPTION":VALUE,"OPTION:"STRING"},"FIELD":{"OPTION":VALUE,"OPTION":"STRING"}}.

You can specify the following faceting options:

buckets specifies an array of the facet values or ranges to count. Ranges are specified using the same syntax that you use to search for a range of values. For more information, see Searching for a Range of Values in the Amazon CloudSearch Developer Guide. Buckets are returned in the order they are specified in the request. The sort and size options are not valid if you specify buckets.

size specifies the maximum number of facets to include in the results. By default, Amazon CloudSearch returns counts for the top 10. The size parameter is only valid when you specify the sort option; it cannot be used in conjunction with buckets.

sort specifies how you want to sort the facets in the results: bucket or count. Specify bucket to sort alphabetically or numerically by facet value (in ascending order). Specify count to sort by the facet counts computed for each facet value (in descending order). To retrieve facet counts for particular values or ranges of values, use the buckets option instead of sort.

If no facet options are specified, facet counts are computed for all field values, the facets are sorted by facet count, and the top 10 facets are returned in the results.

To count particular buckets of values, use the buckets option. For example, the following request uses the buckets option to calculate and return facet counts by decade.

 {"year":{"buckets":["[1970,1979]","[1980,1989]","[1990,1999]","[2000,2009]","[2010,}"]}}

To sort facets by facet count, use the count option. For example, the following request sets the sort option to count to sort the facet values by facet count, with the facet values that have the most matching documents listed first. Setting the size option to 3 returns only the top three facet values.

 {"year":{"sort":"count","size":3}}

To sort the facets by value, use the bucket option. For example, the following request sets the sort option to bucket to sort the facet values numerically by year, with earliest year listed first.

 {"year":{"sort":"bucket"}}

For more information, see Getting and Using Facet Information in the Amazon CloudSearch Developer Guide.

$sel:size:Search', search_size - Specifies the maximum number of search hits to include in the response.

$sel:sort:Search', search_sort - Specifies the fields or custom expressions to use to sort the search results. Multiple fields or expressions are specified as a comma-separated list. You must specify the sort direction (asc or desc) for each field; for example, year desc,title asc. To use a field to sort results, the field must be sort-enabled in the domain configuration. Array type fields cannot be used for sorting. If no sort parameter is specified, results are sorted by their default relevance scores in descending order: _score desc. You can also sort by document ID (_id asc) and version (_version desc).

For more information, see Sorting Results in the Amazon CloudSearch Developer Guide.

$sel:filterQuery:Search', search_filterQuery - Specifies a structured query that filters the results of a search without affecting how the results are scored and sorted. You use filterQuery in conjunction with the query parameter to filter the documents that match the constraints specified in the query parameter. Specifying a filter controls only which matching documents are included in the results, it has no effect on how they are scored and sorted. The filterQuery parameter supports the full structured query syntax.

For more information about using filters, see Filtering Matching Documents in the Amazon CloudSearch Developer Guide.

$sel:query:Search', search_query - Specifies the search criteria for the request. How you specify the search criteria depends on the query parser used for the request and the parser options specified in the queryOptions parameter. By default, the simple query parser is used to process requests. To use the structured, lucene, or dismax query parser, you must also specify the queryParser parameter.

For more information about specifying search criteria, see Searching Your Data in the Amazon CloudSearch Developer Guide.

data SearchResponse

The result of a Search request. Contains the documents that match the specified search criteria and any requested fields, highlights, and facet information.

See: newSearchResponse smart constructor.

Constructors

SearchResponse' (Maybe SearchStatus) (Maybe Hits) (Maybe (HashMap Text FieldStats)) (Maybe (HashMap Text BucketInfo)) Int

newSearchResponse

Arguments

:: Int
$sel:httpStatus:SearchResponse'

-> SearchResponse

Create a value of SearchResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:SearchResponse', searchResponse_status - The status information returned for the search request.

$sel:hits:SearchResponse', searchResponse_hits - The documents that match the search criteria.

$sel:stats:Search', searchResponse_stats - The requested field statistics information.

$sel:facets:SearchResponse', searchResponse_facets - The requested facet information.

$sel:httpStatus:SearchResponse', searchResponse_httpStatus - The response's http status code.

Suggest

data Suggest

Container for the parameters to the Suggest request.

See: newSuggest smart constructor.

Constructors

Suggest' (Maybe Integer) Text Text

newSuggest

Arguments

:: Text
$sel:query:Suggest'

-> Text
$sel:suggester:Suggest'

-> Suggest

Create a value of Suggest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:size:Suggest', suggest_size - Specifies the maximum number of suggestions to return.

$sel:query:Suggest', suggest_query - Specifies the string for which you want to get suggestions.

$sel:suggester:Suggest', suggest_suggester - Specifies the name of the suggester to use to find suggested matches.

data SuggestResponse

Contains the response to a Suggest request.

See: newSuggestResponse smart constructor.

Constructors

SuggestResponse' (Maybe SuggestStatus) (Maybe SuggestModel) Int

newSuggestResponse

Arguments

:: Int
$sel:httpStatus:SuggestResponse'

-> SuggestResponse

Create a value of SuggestResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:SuggestResponse', suggestResponse_status - The status of a SuggestRequest. Contains the resource ID (rid) and how long it took to process the request (timems).

$sel:suggest:SuggestResponse', suggestResponse_suggest - Container for the matching search suggestion information.

$sel:httpStatus:SuggestResponse', suggestResponse_httpStatus - The response's http status code.

Types

ContentType

newtype ContentType

Constructors

ContentType'

Fields

fromContentType :: Text

Bundled Patterns

pattern ContentType_Application_json :: ContentType

pattern ContentType_Application_xml :: ContentType

QueryParser

newtype QueryParser

Constructors

QueryParser'

Fields

fromQueryParser :: Text

Bundled Patterns

pattern QueryParser_Dismax :: QueryParser

pattern QueryParser_Lucene :: QueryParser

pattern QueryParser_Simple :: QueryParser

pattern QueryParser_Structured :: QueryParser

Bucket

data Bucket

A container for facet information.

See: newBucket smart constructor.

Constructors

Bucket' (Maybe Integer) (Maybe Text)

newBucket :: Bucket

Create a value of Bucket with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:count:Bucket', bucket_count - The number of hits that contain the facet value in the specified facet field.

$sel:value:Bucket', bucket_value - The facet value being counted.

BucketInfo

data BucketInfo

A container for the calculated facet values and counts.

See: newBucketInfo smart constructor.

Constructors

BucketInfo' (Maybe [Bucket])

newBucketInfo :: BucketInfo

Create a value of BucketInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:buckets:BucketInfo', bucketInfo_buckets - A list of the calculated facet values and counts.

DocumentServiceWarning

data DocumentServiceWarning

A warning returned by the document service when an issue is discovered while processing an upload request.

See: newDocumentServiceWarning smart constructor.

Constructors

DocumentServiceWarning' (Maybe Text)

newDocumentServiceWarning :: DocumentServiceWarning

Create a value of DocumentServiceWarning with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:message:DocumentServiceWarning', documentServiceWarning_message - The description for a warning returned by the document service.

FieldStats

data FieldStats

The statistics for a field calculated in the request.

See: newFieldStats smart constructor.

Constructors

FieldStats' (Maybe Text) (Maybe Integer) (Maybe Double) (Maybe Text) (Maybe Double) (Maybe Text) (Maybe Integer) (Maybe Double)

newFieldStats :: FieldStats

Create a value of FieldStats with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:mean:FieldStats', fieldStats_mean - The average of the values found in the specified field in the result set.

If the field is numeric (int, int-array, double, or double-array), mean is the string representation of a double-precision 64-bit floating point value. If the field is date or date-array, mean is the string representation of a date with the format specified in IETF RFC3339: yyyy-mm-ddTHH:mm:ss.SSSZ.

$sel:missing:FieldStats', fieldStats_missing - The number of documents that do not contain a value in the specified field in the result set.

$sel:sum:FieldStats', fieldStats_sum - The sum of the field values across the documents in the result set. null for date fields.

$sel:min:FieldStats', fieldStats_min - The minimum value found in the specified field in the result set.

If the field is numeric (int, int-array, double, or double-array), min is the string representation of a double-precision 64-bit floating point value. If the field is date or date-array, min is the string representation of a date with the format specified in IETF RFC3339: yyyy-mm-ddTHH:mm:ss.SSSZ.

$sel:stddev:FieldStats', fieldStats_stddev - The standard deviation of the values in the specified field in the result set.

$sel:max:FieldStats', fieldStats_max - The maximum value found in the specified field in the result set.

If the field is numeric (int, int-array, double, or double-array), max is the string representation of a double-precision 64-bit floating point value. If the field is date or date-array, max is the string representation of a date with the format specified in IETF RFC3339: yyyy-mm-ddTHH:mm:ss.SSSZ.

$sel:count:FieldStats', fieldStats_count - The number of documents that contain a value in the specified field in the result set.

$sel:sumOfSquares:FieldStats', fieldStats_sumOfSquares - The sum of all field values in the result set squared.

Hit

data Hit

Information about a document that matches the search request.

See: newHit smart constructor.

Constructors

Hit' (Maybe Text) (Maybe (HashMap Text Text)) (Maybe (HashMap Text [Text])) (Maybe (HashMap Text Text))

newHit :: Hit

Create a value of Hit with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:Hit', hit_id - The document ID of a document that matches the search request.

$sel:exprs:Hit', hit_exprs - The expressions returned from a document that matches the search request.

$sel:fields:Hit', hit_fields - The fields returned from a document that matches the search request.

$sel:highlights:Hit', hit_highlights - The highlights returned from a document that matches the search request.

Hits

data Hits

The collection of documents that match the search request.

See: newHits smart constructor.

Constructors

Hits' (Maybe Integer) (Maybe [Hit]) (Maybe Text) (Maybe Integer)

newHits :: Hits

Create a value of Hits with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:found:Hits', hits_found - The total number of documents that match the search request.

$sel:hit:Hits', hits_hit - A document that matches the search request.

$sel:cursor:Hits', hits_cursor - A cursor that can be used to retrieve the next set of matching documents when you want to page through a large result set.

$sel:start:Hits', hits_start - The index of the first matching document.

SearchStatus

data SearchStatus

Contains the resource id (rid) and the time it took to process the request (timems).

See: newSearchStatus smart constructor.

Constructors

SearchStatus' (Maybe Integer) (Maybe Text)

newSearchStatus :: SearchStatus

Create a value of SearchStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timems:SearchStatus', searchStatus_timems - How long it took to process the request, in milliseconds.

$sel:rid:SearchStatus', searchStatus_rid - The encrypted resource ID for the request.

SuggestModel

data SuggestModel

Container for the suggestion information returned in a SuggestResponse.

See: newSuggestModel smart constructor.

Constructors

SuggestModel' (Maybe [SuggestionMatch]) (Maybe Integer) (Maybe Text)

newSuggestModel :: SuggestModel

Create a value of SuggestModel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:suggestions:SuggestModel', suggestModel_suggestions - The documents that match the query string.

$sel:found:SuggestModel', suggestModel_found - The number of documents that were found to match the query string.

$sel:query:SuggestModel', suggestModel_query - The query string specified in the suggest request.

SuggestStatus

data SuggestStatus

Contains the resource id (rid) and the time it took to process the request (timems).

See: newSuggestStatus smart constructor.

Constructors

SuggestStatus' (Maybe Integer) (Maybe Text)

newSuggestStatus :: SuggestStatus

Create a value of SuggestStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timems:SuggestStatus', suggestStatus_timems - How long it took to process the request, in milliseconds.

$sel:rid:SuggestStatus', suggestStatus_rid - The encrypted resource ID for the request.

SuggestionMatch

data SuggestionMatch

An autocomplete suggestion that matches the query string specified in a SuggestRequest.

See: newSuggestionMatch smart constructor.

Constructors

SuggestionMatch' (Maybe Text) (Maybe Text) (Maybe Integer)

newSuggestionMatch :: SuggestionMatch

Create a value of SuggestionMatch with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:suggestion:SuggestionMatch', suggestionMatch_suggestion - The string that matches the query string specified in the SuggestRequest.

$sel:id:SuggestionMatch', suggestionMatch_id - The document ID of the suggested document.

$sel:score:SuggestionMatch', suggestionMatch_score - The relevance score of a suggested match.