Network.AWS.CloudSearchDomains.Types

Description

Service Configuration

defaultService :: Service

API version 2013-01-01 of the Amazon CloudSearch Domain SDK configuration.

Errors

_SearchException :: AsError a => Getting (First ServiceError) a ServiceError

Information about any problems encountered while processing a search request.

_DocumentServiceException :: AsError a => Getting (First ServiceError) a ServiceError

Information about any problems encountered while processing an upload request.

ContentType

newtype ContentType

Constructors

ContentType'

Fields

fromContentType :: Text

Bundled Patterns

pattern ContentType_Application_json :: ContentType

pattern ContentType_Application_xml :: ContentType

QueryParser

newtype QueryParser

Constructors

QueryParser'

Fields

fromQueryParser :: Text

Bundled Patterns

pattern QueryParser_Dismax :: QueryParser

pattern QueryParser_Lucene :: QueryParser

pattern QueryParser_Simple :: QueryParser

pattern QueryParser_Structured :: QueryParser

Bucket

data Bucket

A container for facet information.

See: newBucket smart constructor.

Constructors

Bucket'

Fields

count :: Maybe Integer

The number of hits that contain the facet value in the specified facet field.

value :: Maybe Text

The facet value being counted.

newBucket :: Bucket

Create a value of Bucket with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:count:Bucket', bucket_count - The number of hits that contain the facet value in the specified facet field.

$sel:value:Bucket', bucket_value - The facet value being counted.

bucket_count :: Lens' Bucket (Maybe Integer)

The number of hits that contain the facet value in the specified facet field.

bucket_value :: Lens' Bucket (Maybe Text)

The facet value being counted.

BucketInfo

data BucketInfo

A container for the calculated facet values and counts.

See: newBucketInfo smart constructor.

Constructors

BucketInfo'

Fields

buckets :: Maybe [Bucket]

A list of the calculated facet values and counts.

newBucketInfo :: BucketInfo

Create a value of BucketInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:buckets:BucketInfo', bucketInfo_buckets - A list of the calculated facet values and counts.

bucketInfo_buckets :: Lens' BucketInfo (Maybe [Bucket])

A list of the calculated facet values and counts.

DocumentServiceWarning

data DocumentServiceWarning

A warning returned by the document service when an issue is discovered while processing an upload request.

See: newDocumentServiceWarning smart constructor.

Constructors

DocumentServiceWarning'

Fields

message :: Maybe Text

The description for a warning returned by the document service.

newDocumentServiceWarning :: DocumentServiceWarning

Create a value of DocumentServiceWarning with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:message:DocumentServiceWarning', documentServiceWarning_message - The description for a warning returned by the document service.

documentServiceWarning_message :: Lens' DocumentServiceWarning (Maybe Text)

The description for a warning returned by the document service.

FieldStats

data FieldStats

The statistics for a field calculated in the request.

See: newFieldStats smart constructor.

Constructors

FieldStats'

Fields

mean :: Maybe Text

The average of the values found in the specified field in the result set.

If the field is numeric (int, int-array, double, or double-array), mean is the string representation of a double-precision 64-bit floating point value. If the field is date or date-array, mean is the string representation of a date with the format specified in IETF RFC3339: yyyy-mm-ddTHH:mm:ss.SSSZ.

missing :: Maybe Integer

The number of documents that do not contain a value in the specified field in the result set.

sum :: Maybe Double

The sum of the field values across the documents in the result set. null for date fields.

min :: Maybe Text

The minimum value found in the specified field in the result set.

If the field is numeric (int, int-array, double, or double-array), min is the string representation of a double-precision 64-bit floating point value. If the field is date or date-array, min is the string representation of a date with the format specified in IETF RFC3339: yyyy-mm-ddTHH:mm:ss.SSSZ.

stddev :: Maybe Double

The standard deviation of the values in the specified field in the result set.

max :: Maybe Text

The maximum value found in the specified field in the result set.

If the field is numeric (int, int-array, double, or double-array), max is the string representation of a double-precision 64-bit floating point value. If the field is date or date-array, max is the string representation of a date with the format specified in IETF RFC3339: yyyy-mm-ddTHH:mm:ss.SSSZ.

count :: Maybe Integer

The number of documents that contain a value in the specified field in the result set.

sumOfSquares :: Maybe Double

The sum of all field values in the result set squared.

newFieldStats :: FieldStats

Create a value of FieldStats with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:mean:FieldStats', fieldStats_mean - The average of the values found in the specified field in the result set.

If the field is numeric (int, int-array, double, or double-array), mean is the string representation of a double-precision 64-bit floating point value. If the field is date or date-array, mean is the string representation of a date with the format specified in IETF RFC3339: yyyy-mm-ddTHH:mm:ss.SSSZ.

$sel:missing:FieldStats', fieldStats_missing - The number of documents that do not contain a value in the specified field in the result set.

$sel:sum:FieldStats', fieldStats_sum - The sum of the field values across the documents in the result set. null for date fields.

$sel:min:FieldStats', fieldStats_min - The minimum value found in the specified field in the result set.

If the field is numeric (int, int-array, double, or double-array), min is the string representation of a double-precision 64-bit floating point value. If the field is date or date-array, min is the string representation of a date with the format specified in IETF RFC3339: yyyy-mm-ddTHH:mm:ss.SSSZ.

$sel:stddev:FieldStats', fieldStats_stddev - The standard deviation of the values in the specified field in the result set.

$sel:max:FieldStats', fieldStats_max - The maximum value found in the specified field in the result set.

If the field is numeric (int, int-array, double, or double-array), max is the string representation of a double-precision 64-bit floating point value. If the field is date or date-array, max is the string representation of a date with the format specified in IETF RFC3339: yyyy-mm-ddTHH:mm:ss.SSSZ.

$sel:count:FieldStats', fieldStats_count - The number of documents that contain a value in the specified field in the result set.

$sel:sumOfSquares:FieldStats', fieldStats_sumOfSquares - The sum of all field values in the result set squared.

fieldStats_mean :: Lens' FieldStats (Maybe Text)

The average of the values found in the specified field in the result set.

If the field is numeric (int, int-array, double, or double-array), mean is the string representation of a double-precision 64-bit floating point value. If the field is date or date-array, mean is the string representation of a date with the format specified in IETF RFC3339: yyyy-mm-ddTHH:mm:ss.SSSZ.

fieldStats_missing :: Lens' FieldStats (Maybe Integer)

The number of documents that do not contain a value in the specified field in the result set.

fieldStats_sum :: Lens' FieldStats (Maybe Double)

The sum of the field values across the documents in the result set. null for date fields.

fieldStats_min :: Lens' FieldStats (Maybe Text)

The minimum value found in the specified field in the result set.

If the field is numeric (int, int-array, double, or double-array), min is the string representation of a double-precision 64-bit floating point value. If the field is date or date-array, min is the string representation of a date with the format specified in IETF RFC3339: yyyy-mm-ddTHH:mm:ss.SSSZ.

fieldStats_stddev :: Lens' FieldStats (Maybe Double)

The standard deviation of the values in the specified field in the result set.

fieldStats_max :: Lens' FieldStats (Maybe Text)

The maximum value found in the specified field in the result set.

If the field is numeric (int, int-array, double, or double-array), max is the string representation of a double-precision 64-bit floating point value. If the field is date or date-array, max is the string representation of a date with the format specified in IETF RFC3339: yyyy-mm-ddTHH:mm:ss.SSSZ.

fieldStats_count :: Lens' FieldStats (Maybe Integer)

The number of documents that contain a value in the specified field in the result set.

fieldStats_sumOfSquares :: Lens' FieldStats (Maybe Double)

The sum of all field values in the result set squared.

Hit

data Hit

Information about a document that matches the search request.

See: newHit smart constructor.

Constructors

Hit'

Fields

id :: Maybe Text

The document ID of a document that matches the search request.

exprs :: Maybe (HashMap Text Text)

The expressions returned from a document that matches the search request.

fields :: Maybe (HashMap Text [Text])

The fields returned from a document that matches the search request.

highlights :: Maybe (HashMap Text Text)

The highlights returned from a document that matches the search request.

newHit :: Hit

Create a value of Hit with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:Hit', hit_id - The document ID of a document that matches the search request.

$sel:exprs:Hit', hit_exprs - The expressions returned from a document that matches the search request.

$sel:fields:Hit', hit_fields - The fields returned from a document that matches the search request.

$sel:highlights:Hit', hit_highlights - The highlights returned from a document that matches the search request.

hit_id :: Lens' Hit (Maybe Text)

The document ID of a document that matches the search request.

hit_exprs :: Lens' Hit (Maybe (HashMap Text Text))

The expressions returned from a document that matches the search request.

hit_fields :: Lens' Hit (Maybe (HashMap Text [Text]))

The fields returned from a document that matches the search request.

hit_highlights :: Lens' Hit (Maybe (HashMap Text Text))

The highlights returned from a document that matches the search request.

Hits

data Hits

The collection of documents that match the search request.

See: newHits smart constructor.

Constructors

Hits'

Fields

found :: Maybe Integer

The total number of documents that match the search request.

hit :: Maybe [Hit]

A document that matches the search request.

cursor :: Maybe Text

A cursor that can be used to retrieve the next set of matching documents when you want to page through a large result set.

start :: Maybe Integer

The index of the first matching document.

newHits :: Hits

Create a value of Hits with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:found:Hits', hits_found - The total number of documents that match the search request.

$sel:hit:Hits', hits_hit - A document that matches the search request.

$sel:cursor:Hits', hits_cursor - A cursor that can be used to retrieve the next set of matching documents when you want to page through a large result set.

$sel:start:Hits', hits_start - The index of the first matching document.

hits_found :: Lens' Hits (Maybe Integer)

The total number of documents that match the search request.

hits_hit :: Lens' Hits (Maybe [Hit])

A document that matches the search request.

hits_cursor :: Lens' Hits (Maybe Text)

A cursor that can be used to retrieve the next set of matching documents when you want to page through a large result set.

hits_start :: Lens' Hits (Maybe Integer)

The index of the first matching document.

SearchStatus

data SearchStatus

Contains the resource id (rid) and the time it took to process the request (timems).

See: newSearchStatus smart constructor.

Constructors

SearchStatus'

Fields

timems :: Maybe Integer

How long it took to process the request, in milliseconds.

rid :: Maybe Text

The encrypted resource ID for the request.

newSearchStatus :: SearchStatus

Create a value of SearchStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timems:SearchStatus', searchStatus_timems - How long it took to process the request, in milliseconds.

$sel:rid:SearchStatus', searchStatus_rid - The encrypted resource ID for the request.

searchStatus_timems :: Lens' SearchStatus (Maybe Integer)

How long it took to process the request, in milliseconds.

searchStatus_rid :: Lens' SearchStatus (Maybe Text)

The encrypted resource ID for the request.

SuggestModel

data SuggestModel

Container for the suggestion information returned in a SuggestResponse.

See: newSuggestModel smart constructor.

Constructors

SuggestModel'

Fields

suggestions :: Maybe [SuggestionMatch]

The documents that match the query string.

found :: Maybe Integer

The number of documents that were found to match the query string.

query :: Maybe Text

The query string specified in the suggest request.

newSuggestModel :: SuggestModel

Create a value of SuggestModel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:suggestions:SuggestModel', suggestModel_suggestions - The documents that match the query string.

$sel:found:SuggestModel', suggestModel_found - The number of documents that were found to match the query string.

$sel:query:SuggestModel', suggestModel_query - The query string specified in the suggest request.

suggestModel_suggestions :: Lens' SuggestModel (Maybe [SuggestionMatch])

The documents that match the query string.

suggestModel_found :: Lens' SuggestModel (Maybe Integer)

The number of documents that were found to match the query string.

suggestModel_query :: Lens' SuggestModel (Maybe Text)

The query string specified in the suggest request.

SuggestStatus

data SuggestStatus

Contains the resource id (rid) and the time it took to process the request (timems).

See: newSuggestStatus smart constructor.

Constructors

SuggestStatus'

Fields

timems :: Maybe Integer

How long it took to process the request, in milliseconds.

rid :: Maybe Text

The encrypted resource ID for the request.

newSuggestStatus :: SuggestStatus

Create a value of SuggestStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timems:SuggestStatus', suggestStatus_timems - How long it took to process the request, in milliseconds.

$sel:rid:SuggestStatus', suggestStatus_rid - The encrypted resource ID for the request.

suggestStatus_timems :: Lens' SuggestStatus (Maybe Integer)

How long it took to process the request, in milliseconds.

suggestStatus_rid :: Lens' SuggestStatus (Maybe Text)

The encrypted resource ID for the request.

SuggestionMatch

data SuggestionMatch

An autocomplete suggestion that matches the query string specified in a SuggestRequest.

See: newSuggestionMatch smart constructor.

Constructors

SuggestionMatch'

Fields

suggestion :: Maybe Text

The string that matches the query string specified in the SuggestRequest.

id :: Maybe Text

The document ID of the suggested document.

score :: Maybe Integer

The relevance score of a suggested match.

newSuggestionMatch :: SuggestionMatch

Create a value of SuggestionMatch with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:suggestion:SuggestionMatch', suggestionMatch_suggestion - The string that matches the query string specified in the SuggestRequest.

$sel:id:SuggestionMatch', suggestionMatch_id - The document ID of the suggested document.

$sel:score:SuggestionMatch', suggestionMatch_score - The relevance score of a suggested match.

suggestionMatch_suggestion :: Lens' SuggestionMatch (Maybe Text)

The string that matches the query string specified in the SuggestRequest.

suggestionMatch_id :: Lens' SuggestionMatch (Maybe Text)

The document ID of the suggested document.

suggestionMatch_score :: Lens' SuggestionMatch (Maybe Integer)

The relevance score of a suggested match.