Network.AWS.CloudSearchDomains.Types.FieldStats

Description

Documentation

data FieldStats

The statistics for a field calculated in the request.

See: newFieldStats smart constructor.

Constructors

FieldStats'

Fields

mean :: Maybe Text

The average of the values found in the specified field in the result set.

If the field is numeric (int, int-array, double, or double-array), mean is the string representation of a double-precision 64-bit floating point value. If the field is date or date-array, mean is the string representation of a date with the format specified in IETF RFC3339: yyyy-mm-ddTHH:mm:ss.SSSZ.

missing :: Maybe Integer

The number of documents that do not contain a value in the specified field in the result set.

sum :: Maybe Double

The sum of the field values across the documents in the result set. null for date fields.

min :: Maybe Text

The minimum value found in the specified field in the result set.

If the field is numeric (int, int-array, double, or double-array), min is the string representation of a double-precision 64-bit floating point value. If the field is date or date-array, min is the string representation of a date with the format specified in IETF RFC3339: yyyy-mm-ddTHH:mm:ss.SSSZ.

stddev :: Maybe Double

The standard deviation of the values in the specified field in the result set.

max :: Maybe Text

The maximum value found in the specified field in the result set.

If the field is numeric (int, int-array, double, or double-array), max is the string representation of a double-precision 64-bit floating point value. If the field is date or date-array, max is the string representation of a date with the format specified in IETF RFC3339: yyyy-mm-ddTHH:mm:ss.SSSZ.

count :: Maybe Integer

The number of documents that contain a value in the specified field in the result set.

sumOfSquares :: Maybe Double

The sum of all field values in the result set squared.

newFieldStats :: FieldStats

Create a value of FieldStats with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:mean:FieldStats', fieldStats_mean - The average of the values found in the specified field in the result set.

If the field is numeric (int, int-array, double, or double-array), mean is the string representation of a double-precision 64-bit floating point value. If the field is date or date-array, mean is the string representation of a date with the format specified in IETF RFC3339: yyyy-mm-ddTHH:mm:ss.SSSZ.

$sel:missing:FieldStats', fieldStats_missing - The number of documents that do not contain a value in the specified field in the result set.

$sel:sum:FieldStats', fieldStats_sum - The sum of the field values across the documents in the result set. null for date fields.

$sel:min:FieldStats', fieldStats_min - The minimum value found in the specified field in the result set.

If the field is numeric (int, int-array, double, or double-array), min is the string representation of a double-precision 64-bit floating point value. If the field is date or date-array, min is the string representation of a date with the format specified in IETF RFC3339: yyyy-mm-ddTHH:mm:ss.SSSZ.

$sel:stddev:FieldStats', fieldStats_stddev - The standard deviation of the values in the specified field in the result set.

$sel:max:FieldStats', fieldStats_max - The maximum value found in the specified field in the result set.

If the field is numeric (int, int-array, double, or double-array), max is the string representation of a double-precision 64-bit floating point value. If the field is date or date-array, max is the string representation of a date with the format specified in IETF RFC3339: yyyy-mm-ddTHH:mm:ss.SSSZ.

$sel:count:FieldStats', fieldStats_count - The number of documents that contain a value in the specified field in the result set.

$sel:sumOfSquares:FieldStats', fieldStats_sumOfSquares - The sum of all field values in the result set squared.

fieldStats_mean :: Lens' FieldStats (Maybe Text)

The average of the values found in the specified field in the result set.

If the field is numeric (int, int-array, double, or double-array), mean is the string representation of a double-precision 64-bit floating point value. If the field is date or date-array, mean is the string representation of a date with the format specified in IETF RFC3339: yyyy-mm-ddTHH:mm:ss.SSSZ.

fieldStats_missing :: Lens' FieldStats (Maybe Integer)

The number of documents that do not contain a value in the specified field in the result set.

fieldStats_sum :: Lens' FieldStats (Maybe Double)

The sum of the field values across the documents in the result set. null for date fields.

fieldStats_min :: Lens' FieldStats (Maybe Text)

The minimum value found in the specified field in the result set.

If the field is numeric (int, int-array, double, or double-array), min is the string representation of a double-precision 64-bit floating point value. If the field is date or date-array, min is the string representation of a date with the format specified in IETF RFC3339: yyyy-mm-ddTHH:mm:ss.SSSZ.

fieldStats_stddev :: Lens' FieldStats (Maybe Double)

The standard deviation of the values in the specified field in the result set.

fieldStats_max :: Lens' FieldStats (Maybe Text)

The maximum value found in the specified field in the result set.

If the field is numeric (int, int-array, double, or double-array), max is the string representation of a double-precision 64-bit floating point value. If the field is date or date-array, max is the string representation of a date with the format specified in IETF RFC3339: yyyy-mm-ddTHH:mm:ss.SSSZ.

fieldStats_count :: Lens' FieldStats (Maybe Integer)

The number of documents that contain a value in the specified field in the result set.

fieldStats_sumOfSquares :: Lens' FieldStats (Maybe Double)

The sum of all field values in the result set squared.