Network.AWS.CloudSearchDomains.Types.Hits

Description

Documentation

data Hits

The collection of documents that match the search request.

See: newHits smart constructor.

Constructors

Hits'

Fields

found :: Maybe Integer

The total number of documents that match the search request.

hit :: Maybe [Hit]

A document that matches the search request.

cursor :: Maybe Text

A cursor that can be used to retrieve the next set of matching documents when you want to page through a large result set.

start :: Maybe Integer

The index of the first matching document.

newHits :: Hits

Create a value of Hits with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:found:Hits', hits_found - The total number of documents that match the search request.

$sel:hit:Hits', hits_hit - A document that matches the search request.

$sel:cursor:Hits', hits_cursor - A cursor that can be used to retrieve the next set of matching documents when you want to page through a large result set.

$sel:start:Hits', hits_start - The index of the first matching document.

hits_found :: Lens' Hits (Maybe Integer)

The total number of documents that match the search request.

hits_hit :: Lens' Hits (Maybe [Hit])

A document that matches the search request.

hits_cursor :: Lens' Hits (Maybe Text)

A cursor that can be used to retrieve the next set of matching documents when you want to page through a large result set.

hits_start :: Lens' Hits (Maybe Integer)

The index of the first matching document.