Network.AWS.CloudSearchDomains.UploadDocuments

Description

Posts a batch of documents to a search domain for indexing. A document batch is a collection of add and delete operations that represent the documents you want to add, update, or delete from your domain. Batches can be described in either JSON or XML. Each item that you want Amazon CloudSearch to return as a search result (such as a product) is represented as a document. Every document has a unique ID and one or more fields that contain the data that you want to search and return in results. Individual documents cannot contain more than 1 MB of data. The entire batch cannot exceed 5 MB. To get the best possible upload performance, group add and delete operations in batches that are close the 5 MB limit. Submitting a large volume of single-document batches can overload a domain's document service.

The endpoint for submitting UploadDocuments requests is domain-specific. To get the document endpoint for your domain, use the Amazon CloudSearch configuration service DescribeDomains action. A domain's endpoints are also displayed on the domain dashboard in the Amazon CloudSearch console.

For more information about formatting your data for Amazon CloudSearch, see Preparing Your Data in the Amazon CloudSearch Developer Guide. For more information about uploading data for indexing, see Uploading Data in the Amazon CloudSearch Developer Guide.

Creating a Request

data UploadDocuments

Container for the parameters to the UploadDocuments request.

See: newUploadDocuments smart constructor.

Constructors

UploadDocuments'

Fields

contentType :: ContentType

The format of the batch you are uploading. Amazon CloudSearch supports two document batch formats:

application/json

application/xml

documents :: HashedBody

A batch of documents formatted in JSON or HTML.

newUploadDocuments

Arguments

:: ContentType
$sel:contentType:UploadDocuments'

-> HashedBody
$sel:documents:UploadDocuments'

-> UploadDocuments

Create a value of UploadDocuments with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:contentType:UploadDocuments', uploadDocuments_contentType - The format of the batch you are uploading. Amazon CloudSearch supports two document batch formats:

application/json

application/xml

$sel:documents:UploadDocuments', uploadDocuments_documents - A batch of documents formatted in JSON or HTML.

Request Lenses

uploadDocuments_contentType :: Lens' UploadDocuments ContentType

The format of the batch you are uploading. Amazon CloudSearch supports two document batch formats:

application/json

application/xml

uploadDocuments_documents :: Lens' UploadDocuments HashedBody

A batch of documents formatted in JSON or HTML.

Destructuring the Response

data UploadDocumentsResponse

Contains the response to an UploadDocuments request.

See: newUploadDocumentsResponse smart constructor.

Constructors

UploadDocumentsResponse'

Fields

status :: Maybe Text

The status of an UploadDocumentsRequest.

warnings :: Maybe [DocumentServiceWarning]

Any warnings returned by the document service about the documents being uploaded.

deletes :: Maybe Integer

The number of documents that were deleted from the search domain.

adds :: Maybe Integer

The number of documents that were added to the search domain.

httpStatus :: Int

The response's http status code.

newUploadDocumentsResponse

Arguments

:: Int
$sel:httpStatus:UploadDocumentsResponse'

-> UploadDocumentsResponse

Create a value of UploadDocumentsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:UploadDocumentsResponse', uploadDocumentsResponse_status - The status of an UploadDocumentsRequest.

$sel:warnings:UploadDocumentsResponse', uploadDocumentsResponse_warnings - Any warnings returned by the document service about the documents being uploaded.

$sel:deletes:UploadDocumentsResponse', uploadDocumentsResponse_deletes - The number of documents that were deleted from the search domain.

$sel:adds:UploadDocumentsResponse', uploadDocumentsResponse_adds - The number of documents that were added to the search domain.

$sel:httpStatus:UploadDocumentsResponse', uploadDocumentsResponse_httpStatus - The response's http status code.

Response Lenses

uploadDocumentsResponse_status :: Lens' UploadDocumentsResponse (Maybe Text)

The status of an UploadDocumentsRequest.

uploadDocumentsResponse_warnings :: Lens' UploadDocumentsResponse (Maybe [DocumentServiceWarning])

Any warnings returned by the document service about the documents being uploaded.

uploadDocumentsResponse_deletes :: Lens' UploadDocumentsResponse (Maybe Integer)

The number of documents that were deleted from the search domain.

uploadDocumentsResponse_adds :: Lens' UploadDocumentsResponse (Maybe Integer)

The number of documents that were added to the search domain.

uploadDocumentsResponse_httpStatus :: Lens' UploadDocumentsResponse Int

The response's http status code.