Network.AWS.CloudTrail.DescribeTrails

Description

Retrieves settings for one or more trails associated with the current region for your account.

Creating a Request

data DescribeTrails

Returns information about the trail.

See: newDescribeTrails smart constructor.

Constructors

DescribeTrails'

Fields

trailNameList :: Maybe [Text]

Specifies a list of trail names, trail ARNs, or both, of the trails to describe. The format of a trail ARN is:

arn:aws:cloudtrail:us-east-2:123456789012:trail/MyTrail

If an empty list is specified, information for the trail in the current region is returned.

If an empty list is specified and IncludeShadowTrails is false, then information for all trails in the current region is returned.

If an empty list is specified and IncludeShadowTrails is null or true, then information for all trails in the current region and any associated shadow trails in other regions is returned.

If one or more trail names are specified, information is returned only if the names match the names of trails belonging only to the current region. To return information about a trail in another region, you must specify its trail ARN.

includeShadowTrails :: Maybe Bool

Specifies whether to include shadow trails in the response. A shadow trail is the replication in a region of a trail that was created in a different region, or in the case of an organization trail, the replication of an organization trail in member accounts. If you do not include shadow trails, organization trails in a member account and region replication trails will not be returned. The default is true.

newDescribeTrails :: DescribeTrails

Create a value of DescribeTrails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:trailNameList:DescribeTrails', describeTrails_trailNameList - Specifies a list of trail names, trail ARNs, or both, of the trails to describe. The format of a trail ARN is:

arn:aws:cloudtrail:us-east-2:123456789012:trail/MyTrail

If an empty list is specified, information for the trail in the current region is returned.

If an empty list is specified and IncludeShadowTrails is false, then information for all trails in the current region is returned.

If an empty list is specified and IncludeShadowTrails is null or true, then information for all trails in the current region and any associated shadow trails in other regions is returned.

If one or more trail names are specified, information is returned only if the names match the names of trails belonging only to the current region. To return information about a trail in another region, you must specify its trail ARN.

$sel:includeShadowTrails:DescribeTrails', describeTrails_includeShadowTrails - Specifies whether to include shadow trails in the response. A shadow trail is the replication in a region of a trail that was created in a different region, or in the case of an organization trail, the replication of an organization trail in member accounts. If you do not include shadow trails, organization trails in a member account and region replication trails will not be returned. The default is true.

Request Lenses

describeTrails_trailNameList :: Lens' DescribeTrails (Maybe [Text])

Specifies a list of trail names, trail ARNs, or both, of the trails to describe. The format of a trail ARN is:

arn:aws:cloudtrail:us-east-2:123456789012:trail/MyTrail

If an empty list is specified, information for the trail in the current region is returned.

If an empty list is specified and IncludeShadowTrails is false, then information for all trails in the current region is returned.

If an empty list is specified and IncludeShadowTrails is null or true, then information for all trails in the current region and any associated shadow trails in other regions is returned.

If one or more trail names are specified, information is returned only if the names match the names of trails belonging only to the current region. To return information about a trail in another region, you must specify its trail ARN.

describeTrails_includeShadowTrails :: Lens' DescribeTrails (Maybe Bool)

Specifies whether to include shadow trails in the response. A shadow trail is the replication in a region of a trail that was created in a different region, or in the case of an organization trail, the replication of an organization trail in member accounts. If you do not include shadow trails, organization trails in a member account and region replication trails will not be returned. The default is true.

Destructuring the Response

data DescribeTrailsResponse

Returns the objects or data listed below if successful. Otherwise, returns an error.

See: newDescribeTrailsResponse smart constructor.

Constructors

DescribeTrailsResponse'

Fields

trailList :: Maybe [Trail]

The list of trail objects. Trail objects with string values are only returned if values for the objects exist in a trail's configuration. For example, SNSTopicName and SNSTopicARN are only returned in results if a trail is configured to send SNS notifications. Similarly, KMSKeyId only appears in results if a trail's log files are encrypted with KMS customer managed keys.

httpStatus :: Int

The response's http status code.

newDescribeTrailsResponse

Arguments

:: Int
$sel:httpStatus:DescribeTrailsResponse'

-> DescribeTrailsResponse

Create a value of DescribeTrailsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:trailList:DescribeTrailsResponse', describeTrailsResponse_trailList - The list of trail objects. Trail objects with string values are only returned if values for the objects exist in a trail's configuration. For example, SNSTopicName and SNSTopicARN are only returned in results if a trail is configured to send SNS notifications. Similarly, KMSKeyId only appears in results if a trail's log files are encrypted with KMS customer managed keys.

$sel:httpStatus:DescribeTrailsResponse', describeTrailsResponse_httpStatus - The response's http status code.

Response Lenses

describeTrailsResponse_trailList :: Lens' DescribeTrailsResponse (Maybe [Trail])

The list of trail objects. Trail objects with string values are only returned if values for the objects exist in a trail's configuration. For example, SNSTopicName and SNSTopicARN are only returned in results if a trail is configured to send SNS notifications. Similarly, KMSKeyId only appears in results if a trail's log files are encrypted with KMS customer managed keys.

describeTrailsResponse_httpStatus :: Lens' DescribeTrailsResponse Int

The response's http status code.