Network.AWS.CloudTrail.GetEventSelectors

Description

Describes the settings for the event selectors that you configured for your trail. The information returned for your event selectors includes the following:

If your event selector includes read-only events, write-only events, or all events. This applies to both management events and data events.

If your event selector includes management events.

If your event selector includes data events, the resources on which you are logging data events.

For more information, see Logging Data and Management Events for Trails in the CloudTrail User Guide.

Creating a Request

data GetEventSelectors

See: newGetEventSelectors smart constructor.

Constructors

GetEventSelectors'

Fields

trailName :: Text

Specifies the name of the trail or trail ARN. If you specify a trail name, the string must meet the following requirements:

Contain only ASCII letters (a-z, A-Z), numbers (0-9), periods (.), underscores (_), or dashes (-)

Start with a letter or number, and end with a letter or number

Be between 3 and 128 characters

Have no adjacent periods, underscores or dashes. Names like my-_namespace and my--namespace are not valid.

Not be in IP address format (for example, 192.168.5.4)

If you specify a trail ARN, it must be in the format:

arn:aws:cloudtrail:us-east-2:123456789012:trail/MyTrail

newGetEventSelectors

Arguments

:: Text
$sel:trailName:GetEventSelectors'

-> GetEventSelectors

Create a value of GetEventSelectors with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:trailName:GetEventSelectors', getEventSelectors_trailName - Specifies the name of the trail or trail ARN. If you specify a trail name, the string must meet the following requirements:

Contain only ASCII letters (a-z, A-Z), numbers (0-9), periods (.), underscores (_), or dashes (-)

Start with a letter or number, and end with a letter or number

Be between 3 and 128 characters

Have no adjacent periods, underscores or dashes. Names like my-_namespace and my--namespace are not valid.

Not be in IP address format (for example, 192.168.5.4)

If you specify a trail ARN, it must be in the format:

arn:aws:cloudtrail:us-east-2:123456789012:trail/MyTrail

Request Lenses

getEventSelectors_trailName :: Lens' GetEventSelectors Text

Specifies the name of the trail or trail ARN. If you specify a trail name, the string must meet the following requirements:

Contain only ASCII letters (a-z, A-Z), numbers (0-9), periods (.), underscores (_), or dashes (-)

Start with a letter or number, and end with a letter or number

Be between 3 and 128 characters

Have no adjacent periods, underscores or dashes. Names like my-_namespace and my--namespace are not valid.

Not be in IP address format (for example, 192.168.5.4)

If you specify a trail ARN, it must be in the format:

arn:aws:cloudtrail:us-east-2:123456789012:trail/MyTrail

Destructuring the Response

data GetEventSelectorsResponse

See: newGetEventSelectorsResponse smart constructor.

Constructors

GetEventSelectorsResponse'

Fields

trailARN :: Maybe Text

The specified trail ARN that has the event selectors.

eventSelectors :: Maybe [EventSelector]

The event selectors that are configured for the trail.

advancedEventSelectors :: Maybe [AdvancedEventSelector]

The advanced event selectors that are configured for the trail.

httpStatus :: Int

The response's http status code.

newGetEventSelectorsResponse

Arguments

:: Int
$sel:httpStatus:GetEventSelectorsResponse'

-> GetEventSelectorsResponse

Create a value of GetEventSelectorsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:trailARN:GetEventSelectorsResponse', getEventSelectorsResponse_trailARN - The specified trail ARN that has the event selectors.

$sel:eventSelectors:GetEventSelectorsResponse', getEventSelectorsResponse_eventSelectors - The event selectors that are configured for the trail.

$sel:advancedEventSelectors:GetEventSelectorsResponse', getEventSelectorsResponse_advancedEventSelectors - The advanced event selectors that are configured for the trail.

$sel:httpStatus:GetEventSelectorsResponse', getEventSelectorsResponse_httpStatus - The response's http status code.

Response Lenses

getEventSelectorsResponse_trailARN :: Lens' GetEventSelectorsResponse (Maybe Text)

The specified trail ARN that has the event selectors.

getEventSelectorsResponse_eventSelectors :: Lens' GetEventSelectorsResponse (Maybe [EventSelector])

The event selectors that are configured for the trail.

getEventSelectorsResponse_advancedEventSelectors :: Lens' GetEventSelectorsResponse (Maybe [AdvancedEventSelector])

The advanced event selectors that are configured for the trail.

getEventSelectorsResponse_httpStatus :: Lens' GetEventSelectorsResponse Int

The response's http status code.