Network.AWS.CloudTrail.GetInsightSelectors

Description

Describes the settings for the Insights event selectors that you configured for your trail. GetInsightSelectors shows if CloudTrail Insights event logging is enabled on the trail, and if it is, which insight types are enabled. If you run GetInsightSelectors on a trail that does not have Insights events enabled, the operation throws the exception InsightNotEnabledException

For more information, see Logging CloudTrail Insights Events for Trails in the CloudTrail User Guide.

Creating a Request

data GetInsightSelectors

See: newGetInsightSelectors smart constructor.

Constructors

GetInsightSelectors'

Fields

trailName :: Text

Specifies the name of the trail or trail ARN. If you specify a trail name, the string must meet the following requirements:

Contain only ASCII letters (a-z, A-Z), numbers (0-9), periods (.), underscores (_), or dashes (-)

Start with a letter or number, and end with a letter or number

Be between 3 and 128 characters

Have no adjacent periods, underscores or dashes. Names like my-_namespace and my--namespace are not valid.

Not be in IP address format (for example, 192.168.5.4)

If you specify a trail ARN, it must be in the format:

arn:aws:cloudtrail:us-east-2:123456789012:trail/MyTrail

newGetInsightSelectors

Arguments

:: Text
$sel:trailName:GetInsightSelectors'

-> GetInsightSelectors

Create a value of GetInsightSelectors with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:trailName:GetInsightSelectors', getInsightSelectors_trailName - Specifies the name of the trail or trail ARN. If you specify a trail name, the string must meet the following requirements:

Contain only ASCII letters (a-z, A-Z), numbers (0-9), periods (.), underscores (_), or dashes (-)

Start with a letter or number, and end with a letter or number

Be between 3 and 128 characters

Have no adjacent periods, underscores or dashes. Names like my-_namespace and my--namespace are not valid.

Not be in IP address format (for example, 192.168.5.4)

If you specify a trail ARN, it must be in the format:

arn:aws:cloudtrail:us-east-2:123456789012:trail/MyTrail

Request Lenses

getInsightSelectors_trailName :: Lens' GetInsightSelectors Text

Specifies the name of the trail or trail ARN. If you specify a trail name, the string must meet the following requirements:

Contain only ASCII letters (a-z, A-Z), numbers (0-9), periods (.), underscores (_), or dashes (-)

Start with a letter or number, and end with a letter or number

Be between 3 and 128 characters

Have no adjacent periods, underscores or dashes. Names like my-_namespace and my--namespace are not valid.

Not be in IP address format (for example, 192.168.5.4)

If you specify a trail ARN, it must be in the format:

arn:aws:cloudtrail:us-east-2:123456789012:trail/MyTrail

Destructuring the Response

data GetInsightSelectorsResponse

See: newGetInsightSelectorsResponse smart constructor.

Constructors

GetInsightSelectorsResponse'

Fields

trailARN :: Maybe Text

The Amazon Resource Name (ARN) of a trail for which you want to get Insights selectors.

insightSelectors :: Maybe [InsightSelector]

A JSON string that contains the insight types you want to log on a trail. In this release, only ApiCallRateInsight is supported as an insight type.

httpStatus :: Int

The response's http status code.

newGetInsightSelectorsResponse

Arguments

:: Int
$sel:httpStatus:GetInsightSelectorsResponse'

-> GetInsightSelectorsResponse

Create a value of GetInsightSelectorsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:trailARN:GetInsightSelectorsResponse', getInsightSelectorsResponse_trailARN - The Amazon Resource Name (ARN) of a trail for which you want to get Insights selectors.

$sel:insightSelectors:GetInsightSelectorsResponse', getInsightSelectorsResponse_insightSelectors - A JSON string that contains the insight types you want to log on a trail. In this release, only ApiCallRateInsight is supported as an insight type.

$sel:httpStatus:GetInsightSelectorsResponse', getInsightSelectorsResponse_httpStatus - The response's http status code.

Response Lenses

getInsightSelectorsResponse_trailARN :: Lens' GetInsightSelectorsResponse (Maybe Text)

The Amazon Resource Name (ARN) of a trail for which you want to get Insights selectors.

getInsightSelectorsResponse_insightSelectors :: Lens' GetInsightSelectorsResponse (Maybe [InsightSelector])

A JSON string that contains the insight types you want to log on a trail. In this release, only ApiCallRateInsight is supported as an insight type.

getInsightSelectorsResponse_httpStatus :: Lens' GetInsightSelectorsResponse Int

The response's http status code.