Network.AWS.CloudTrail.GetTrailStatus

Description

Returns a JSON-formatted list of information about the specified trail. Fields include information on delivery errors, Amazon SNS and Amazon S3 errors, and start and stop logging times for each trail. This operation returns trail status from a single region. To return trail status from all regions, you must call the operation on each region.

Creating a Request

data GetTrailStatus

The name of a trail about which you want the current status.

See: newGetTrailStatus smart constructor.

Constructors

GetTrailStatus'

Fields

name :: Text

Specifies the name or the CloudTrail ARN of the trail for which you are requesting status. To get the status of a shadow trail (a replication of the trail in another region), you must specify its ARN. The following is the format of a trail ARN.

arn:aws:cloudtrail:us-east-2:123456789012:trail/MyTrail

newGetTrailStatus

Arguments

:: Text
$sel:name:GetTrailStatus'

-> GetTrailStatus

Create a value of GetTrailStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:GetTrailStatus', getTrailStatus_name - Specifies the name or the CloudTrail ARN of the trail for which you are requesting status. To get the status of a shadow trail (a replication of the trail in another region), you must specify its ARN. The following is the format of a trail ARN.

arn:aws:cloudtrail:us-east-2:123456789012:trail/MyTrail

Request Lenses

getTrailStatus_name :: Lens' GetTrailStatus Text

Specifies the name or the CloudTrail ARN of the trail for which you are requesting status. To get the status of a shadow trail (a replication of the trail in another region), you must specify its ARN. The following is the format of a trail ARN.

arn:aws:cloudtrail:us-east-2:123456789012:trail/MyTrail

Destructuring the Response

data GetTrailStatusResponse

Returns the objects or data listed below if successful. Otherwise, returns an error.

See: newGetTrailStatusResponse smart constructor.

Constructors

GetTrailStatusResponse'

Fields

timeLoggingStopped :: Maybe Text

This field is no longer in use.

latestDeliveryAttemptTime :: Maybe Text

This field is no longer in use.

latestDigestDeliveryError :: Maybe Text

Displays any Amazon S3 error that CloudTrail encountered when attempting to deliver a digest file to the designated bucket. For more information, see Error Responses in the Amazon S3 API Reference.

This error occurs only when there is a problem with the destination S3 bucket, and does not occur for requests that time out. To resolve the issue, create a new bucket, and then call UpdateTrail to specify the new bucket; or fix the existing objects so that CloudTrail can again write to the bucket.

latestDeliveryAttemptSucceeded :: Maybe Text

This field is no longer in use.

latestNotificationError :: Maybe Text

Displays any Amazon SNS error that CloudTrail encountered when attempting to send a notification. For more information about Amazon SNS errors, see the Amazon SNS Developer Guide.

latestCloudWatchLogsDeliveryError :: Maybe Text

Displays any CloudWatch Logs error that CloudTrail encountered when attempting to deliver logs to CloudWatch Logs.

latestNotificationAttemptSucceeded :: Maybe Text

This field is no longer in use.

latestCloudWatchLogsDeliveryTime :: Maybe POSIX

Displays the most recent date and time when CloudTrail delivered logs to CloudWatch Logs.

latestDigestDeliveryTime :: Maybe POSIX

Specifies the date and time that CloudTrail last delivered a digest file to an account's Amazon S3 bucket.

timeLoggingStarted :: Maybe Text

This field is no longer in use.

latestDeliveryError :: Maybe Text

Displays any Amazon S3 error that CloudTrail encountered when attempting to deliver log files to the designated bucket. For more information, see Error Responses in the Amazon S3 API Reference.

This error occurs only when there is a problem with the destination S3 bucket, and does not occur for requests that time out. To resolve the issue, create a new bucket, and then call UpdateTrail to specify the new bucket; or fix the existing objects so that CloudTrail can again write to the bucket.

isLogging :: Maybe Bool

Whether the CloudTrail trail is currently logging Amazon Web Services API calls.

latestNotificationAttemptTime :: Maybe Text

This field is no longer in use.

startLoggingTime :: Maybe POSIX

Specifies the most recent date and time when CloudTrail started recording API calls for an Amazon Web Services account.

stopLoggingTime :: Maybe POSIX

Specifies the most recent date and time when CloudTrail stopped recording API calls for an Amazon Web Services account.

latestNotificationTime :: Maybe POSIX

Specifies the date and time of the most recent Amazon SNS notification that CloudTrail has written a new log file to an account's Amazon S3 bucket.

latestDeliveryTime :: Maybe POSIX

Specifies the date and time that CloudTrail last delivered log files to an account's Amazon S3 bucket.

httpStatus :: Int

The response's http status code.

newGetTrailStatusResponse

Arguments

:: Int
$sel:httpStatus:GetTrailStatusResponse'

-> GetTrailStatusResponse

Create a value of GetTrailStatusResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timeLoggingStopped:GetTrailStatusResponse', getTrailStatusResponse_timeLoggingStopped - This field is no longer in use.

$sel:latestDeliveryAttemptTime:GetTrailStatusResponse', getTrailStatusResponse_latestDeliveryAttemptTime - This field is no longer in use.

$sel:latestDigestDeliveryError:GetTrailStatusResponse', getTrailStatusResponse_latestDigestDeliveryError - Displays any Amazon S3 error that CloudTrail encountered when attempting to deliver a digest file to the designated bucket. For more information, see Error Responses in the Amazon S3 API Reference.

This error occurs only when there is a problem with the destination S3 bucket, and does not occur for requests that time out. To resolve the issue, create a new bucket, and then call UpdateTrail to specify the new bucket; or fix the existing objects so that CloudTrail can again write to the bucket.

$sel:latestDeliveryAttemptSucceeded:GetTrailStatusResponse', getTrailStatusResponse_latestDeliveryAttemptSucceeded - This field is no longer in use.

$sel:latestNotificationError:GetTrailStatusResponse', getTrailStatusResponse_latestNotificationError - Displays any Amazon SNS error that CloudTrail encountered when attempting to send a notification. For more information about Amazon SNS errors, see the Amazon SNS Developer Guide.

$sel:latestCloudWatchLogsDeliveryError:GetTrailStatusResponse', getTrailStatusResponse_latestCloudWatchLogsDeliveryError - Displays any CloudWatch Logs error that CloudTrail encountered when attempting to deliver logs to CloudWatch Logs.

$sel:latestNotificationAttemptSucceeded:GetTrailStatusResponse', getTrailStatusResponse_latestNotificationAttemptSucceeded - This field is no longer in use.

$sel:latestCloudWatchLogsDeliveryTime:GetTrailStatusResponse', getTrailStatusResponse_latestCloudWatchLogsDeliveryTime - Displays the most recent date and time when CloudTrail delivered logs to CloudWatch Logs.

$sel:latestDigestDeliveryTime:GetTrailStatusResponse', getTrailStatusResponse_latestDigestDeliveryTime - Specifies the date and time that CloudTrail last delivered a digest file to an account's Amazon S3 bucket.

$sel:timeLoggingStarted:GetTrailStatusResponse', getTrailStatusResponse_timeLoggingStarted - This field is no longer in use.

$sel:latestDeliveryError:GetTrailStatusResponse', getTrailStatusResponse_latestDeliveryError - Displays any Amazon S3 error that CloudTrail encountered when attempting to deliver log files to the designated bucket. For more information, see Error Responses in the Amazon S3 API Reference.

This error occurs only when there is a problem with the destination S3 bucket, and does not occur for requests that time out. To resolve the issue, create a new bucket, and then call UpdateTrail to specify the new bucket; or fix the existing objects so that CloudTrail can again write to the bucket.

$sel:isLogging:GetTrailStatusResponse', getTrailStatusResponse_isLogging - Whether the CloudTrail trail is currently logging Amazon Web Services API calls.

$sel:latestNotificationAttemptTime:GetTrailStatusResponse', getTrailStatusResponse_latestNotificationAttemptTime - This field is no longer in use.

$sel:startLoggingTime:GetTrailStatusResponse', getTrailStatusResponse_startLoggingTime - Specifies the most recent date and time when CloudTrail started recording API calls for an Amazon Web Services account.

$sel:stopLoggingTime:GetTrailStatusResponse', getTrailStatusResponse_stopLoggingTime - Specifies the most recent date and time when CloudTrail stopped recording API calls for an Amazon Web Services account.

$sel:latestNotificationTime:GetTrailStatusResponse', getTrailStatusResponse_latestNotificationTime - Specifies the date and time of the most recent Amazon SNS notification that CloudTrail has written a new log file to an account's Amazon S3 bucket.

$sel:latestDeliveryTime:GetTrailStatusResponse', getTrailStatusResponse_latestDeliveryTime - Specifies the date and time that CloudTrail last delivered log files to an account's Amazon S3 bucket.

$sel:httpStatus:GetTrailStatusResponse', getTrailStatusResponse_httpStatus - The response's http status code.

Response Lenses

getTrailStatusResponse_timeLoggingStopped :: Lens' GetTrailStatusResponse (Maybe Text)

This field is no longer in use.

getTrailStatusResponse_latestDeliveryAttemptTime :: Lens' GetTrailStatusResponse (Maybe Text)

This field is no longer in use.

getTrailStatusResponse_latestDigestDeliveryError :: Lens' GetTrailStatusResponse (Maybe Text)

Displays any Amazon S3 error that CloudTrail encountered when attempting to deliver a digest file to the designated bucket. For more information, see Error Responses in the Amazon S3 API Reference.

This error occurs only when there is a problem with the destination S3 bucket, and does not occur for requests that time out. To resolve the issue, create a new bucket, and then call UpdateTrail to specify the new bucket; or fix the existing objects so that CloudTrail can again write to the bucket.

getTrailStatusResponse_latestDeliveryAttemptSucceeded :: Lens' GetTrailStatusResponse (Maybe Text)

This field is no longer in use.

getTrailStatusResponse_latestNotificationError :: Lens' GetTrailStatusResponse (Maybe Text)

Displays any Amazon SNS error that CloudTrail encountered when attempting to send a notification. For more information about Amazon SNS errors, see the Amazon SNS Developer Guide.

getTrailStatusResponse_latestCloudWatchLogsDeliveryError :: Lens' GetTrailStatusResponse (Maybe Text)

Displays any CloudWatch Logs error that CloudTrail encountered when attempting to deliver logs to CloudWatch Logs.

getTrailStatusResponse_latestNotificationAttemptSucceeded :: Lens' GetTrailStatusResponse (Maybe Text)

This field is no longer in use.

getTrailStatusResponse_latestCloudWatchLogsDeliveryTime :: Lens' GetTrailStatusResponse (Maybe UTCTime)

Displays the most recent date and time when CloudTrail delivered logs to CloudWatch Logs.

getTrailStatusResponse_latestDigestDeliveryTime :: Lens' GetTrailStatusResponse (Maybe UTCTime)

Specifies the date and time that CloudTrail last delivered a digest file to an account's Amazon S3 bucket.

getTrailStatusResponse_timeLoggingStarted :: Lens' GetTrailStatusResponse (Maybe Text)

This field is no longer in use.

getTrailStatusResponse_latestDeliveryError :: Lens' GetTrailStatusResponse (Maybe Text)

Displays any Amazon S3 error that CloudTrail encountered when attempting to deliver log files to the designated bucket. For more information, see Error Responses in the Amazon S3 API Reference.

This error occurs only when there is a problem with the destination S3 bucket, and does not occur for requests that time out. To resolve the issue, create a new bucket, and then call UpdateTrail to specify the new bucket; or fix the existing objects so that CloudTrail can again write to the bucket.

getTrailStatusResponse_isLogging :: Lens' GetTrailStatusResponse (Maybe Bool)

Whether the CloudTrail trail is currently logging Amazon Web Services API calls.

getTrailStatusResponse_latestNotificationAttemptTime :: Lens' GetTrailStatusResponse (Maybe Text)

This field is no longer in use.

getTrailStatusResponse_startLoggingTime :: Lens' GetTrailStatusResponse (Maybe UTCTime)

Specifies the most recent date and time when CloudTrail started recording API calls for an Amazon Web Services account.

getTrailStatusResponse_stopLoggingTime :: Lens' GetTrailStatusResponse (Maybe UTCTime)

Specifies the most recent date and time when CloudTrail stopped recording API calls for an Amazon Web Services account.

getTrailStatusResponse_latestNotificationTime :: Lens' GetTrailStatusResponse (Maybe UTCTime)

Specifies the date and time of the most recent Amazon SNS notification that CloudTrail has written a new log file to an account's Amazon S3 bucket.

getTrailStatusResponse_latestDeliveryTime :: Lens' GetTrailStatusResponse (Maybe UTCTime)

Specifies the date and time that CloudTrail last delivered log files to an account's Amazon S3 bucket.

getTrailStatusResponse_httpStatus :: Lens' GetTrailStatusResponse Int

The response's http status code.