Network.AWS.CloudTrail.ListPublicKeys

Description

Returns all public keys whose private keys were used to sign the digest files within the specified time range. The public key is needed to validate digest files that were signed with its corresponding private key.

CloudTrail uses different private and public key pairs per region. Each digest file is signed with a private key unique to its region. When you validate a digest file from a specific region, you must look in the same region for its corresponding public key.

This operation returns paginated results.

Creating a Request

data ListPublicKeys

Requests the public keys for a specified time range.

See: newListPublicKeys smart constructor.

Constructors

ListPublicKeys'

Fields

startTime :: Maybe POSIX

Optionally specifies, in UTC, the start of the time range to look up public keys for CloudTrail digest files. If not specified, the current time is used, and the current public key is returned.

nextToken :: Maybe Text

Reserved for future use.

endTime :: Maybe POSIX

Optionally specifies, in UTC, the end of the time range to look up public keys for CloudTrail digest files. If not specified, the current time is used.

newListPublicKeys :: ListPublicKeys

Create a value of ListPublicKeys with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:startTime:ListPublicKeys', listPublicKeys_startTime - Optionally specifies, in UTC, the start of the time range to look up public keys for CloudTrail digest files. If not specified, the current time is used, and the current public key is returned.

$sel:nextToken:ListPublicKeys', listPublicKeys_nextToken - Reserved for future use.

$sel:endTime:ListPublicKeys', listPublicKeys_endTime - Optionally specifies, in UTC, the end of the time range to look up public keys for CloudTrail digest files. If not specified, the current time is used.

Request Lenses

listPublicKeys_startTime :: Lens' ListPublicKeys (Maybe UTCTime)

Optionally specifies, in UTC, the start of the time range to look up public keys for CloudTrail digest files. If not specified, the current time is used, and the current public key is returned.

listPublicKeys_nextToken :: Lens' ListPublicKeys (Maybe Text)

Reserved for future use.

listPublicKeys_endTime :: Lens' ListPublicKeys (Maybe UTCTime)

Optionally specifies, in UTC, the end of the time range to look up public keys for CloudTrail digest files. If not specified, the current time is used.

Destructuring the Response

data ListPublicKeysResponse

Returns the objects or data listed below if successful. Otherwise, returns an error.

See: newListPublicKeysResponse smart constructor.

Constructors

ListPublicKeysResponse'

Fields

publicKeyList :: Maybe [PublicKey]

Contains an array of PublicKey objects.

The returned public keys may have validity time ranges that overlap.

nextToken :: Maybe Text

Reserved for future use.

httpStatus :: Int

The response's http status code.

newListPublicKeysResponse

Arguments

:: Int
$sel:httpStatus:ListPublicKeysResponse'

-> ListPublicKeysResponse

Create a value of ListPublicKeysResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:publicKeyList:ListPublicKeysResponse', listPublicKeysResponse_publicKeyList - Contains an array of PublicKey objects.

The returned public keys may have validity time ranges that overlap.

$sel:nextToken:ListPublicKeys', listPublicKeysResponse_nextToken - Reserved for future use.

$sel:httpStatus:ListPublicKeysResponse', listPublicKeysResponse_httpStatus - The response's http status code.

Response Lenses

listPublicKeysResponse_publicKeyList :: Lens' ListPublicKeysResponse (Maybe [PublicKey])

Contains an array of PublicKey objects.

The returned public keys may have validity time ranges that overlap.

listPublicKeysResponse_nextToken :: Lens' ListPublicKeysResponse (Maybe Text)

Reserved for future use.

listPublicKeysResponse_httpStatus :: Lens' ListPublicKeysResponse Int

The response's http status code.