Network.AWS.CloudTrail.ListTrails

Description

Lists trails that are in the current account.

This operation returns paginated results.

Creating a Request

data ListTrails

See: newListTrails smart constructor.

Constructors

ListTrails'

Fields

nextToken :: Maybe Text

The token to use to get the next page of results after a previous API call. This token must be passed in with the same parameters that were specified in the the original call. For example, if the original call specified an AttributeKey of 'Username' with a value of 'root', the call with NextToken should include those same parameters.

newListTrails :: ListTrails

Create a value of ListTrails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListTrails', listTrails_nextToken - The token to use to get the next page of results after a previous API call. This token must be passed in with the same parameters that were specified in the the original call. For example, if the original call specified an AttributeKey of 'Username' with a value of 'root', the call with NextToken should include those same parameters.

Request Lenses

listTrails_nextToken :: Lens' ListTrails (Maybe Text)

The token to use to get the next page of results after a previous API call. This token must be passed in with the same parameters that were specified in the the original call. For example, if the original call specified an AttributeKey of 'Username' with a value of 'root', the call with NextToken should include those same parameters.

Destructuring the Response

data ListTrailsResponse

See: newListTrailsResponse smart constructor.

Constructors

ListTrailsResponse'

Fields

nextToken :: Maybe Text

The token to use to get the next page of results after a previous API call. If the token does not appear, there are no more results to return. The token must be passed in with the same parameters as the previous call. For example, if the original call specified an AttributeKey of 'Username' with a value of 'root', the call with NextToken should include those same parameters.

trails :: Maybe [TrailInfo]

Returns the name, ARN, and home region of trails in the current account.

httpStatus :: Int

The response's http status code.

newListTrailsResponse

Arguments

:: Int
$sel:httpStatus:ListTrailsResponse'

-> ListTrailsResponse

Create a value of ListTrailsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListTrails', listTrailsResponse_nextToken - The token to use to get the next page of results after a previous API call. If the token does not appear, there are no more results to return. The token must be passed in with the same parameters as the previous call. For example, if the original call specified an AttributeKey of 'Username' with a value of 'root', the call with NextToken should include those same parameters.

$sel:trails:ListTrailsResponse', listTrailsResponse_trails - Returns the name, ARN, and home region of trails in the current account.

$sel:httpStatus:ListTrailsResponse', listTrailsResponse_httpStatus - The response's http status code.

Response Lenses

listTrailsResponse_nextToken :: Lens' ListTrailsResponse (Maybe Text)

The token to use to get the next page of results after a previous API call. If the token does not appear, there are no more results to return. The token must be passed in with the same parameters as the previous call. For example, if the original call specified an AttributeKey of 'Username' with a value of 'root', the call with NextToken should include those same parameters.

listTrailsResponse_trails :: Lens' ListTrailsResponse (Maybe [TrailInfo])

Returns the name, ARN, and home region of trails in the current account.

listTrailsResponse_httpStatus :: Lens' ListTrailsResponse Int

The response's http status code.