Network.AWS.CloudTrail.PutEventSelectors

Description

Configures an event selector or advanced event selectors for your trail. Use event selectors or advanced event selectors to specify management and data event settings for your trail. By default, trails created without specific event selectors are configured to log all read and write management events, and no data events.

When an event occurs in your account, CloudTrail evaluates the event selectors or advanced event selectors in all trails. For each trail, if the event matches any event selector, the trail processes and logs the event. If the event doesn't match any event selector, the trail doesn't log the event.

Example

You create an event selector for a trail and specify that you want write-only events.

The EC2 GetConsoleOutput and RunInstances API operations occur in your account.

CloudTrail evaluates whether the events match your event selectors.

The RunInstances is a write-only event and it matches your event selector. The trail logs the event.

The GetConsoleOutput is a read-only event that doesn't match your event selector. The trail doesn't log the event.

The PutEventSelectors operation must be called from the region in which the trail was created; otherwise, an InvalidHomeRegionException exception is thrown.

You can configure up to five event selectors for each trail. For more information, see Logging data and management events for trails and Quotas in CloudTrail in the CloudTrail User Guide.

You can add advanced event selectors, and conditions for your advanced event selectors, up to a maximum of 500 values for all conditions and selectors on a trail. You can use either AdvancedEventSelectors or EventSelectors, but not both. If you apply AdvancedEventSelectors to a trail, any existing EventSelectors are overwritten. For more information about advanced event selectors, see Logging data events for trails in the CloudTrail User Guide.

Creating a Request

data PutEventSelectors

See: newPutEventSelectors smart constructor.

Constructors

PutEventSelectors'

Fields

eventSelectors :: Maybe [EventSelector]

Specifies the settings for your event selectors. You can configure up to five event selectors for a trail. You can use either EventSelectors or AdvancedEventSelectors in a PutEventSelectors request, but not both. If you apply EventSelectors to a trail, any existing AdvancedEventSelectors are overwritten.

advancedEventSelectors :: Maybe [AdvancedEventSelector]

Specifies the settings for advanced event selectors. You can add advanced event selectors, and conditions for your advanced event selectors, up to a maximum of 500 values for all conditions and selectors on a trail. You can use either AdvancedEventSelectors or EventSelectors, but not both. If you apply AdvancedEventSelectors to a trail, any existing EventSelectors are overwritten. For more information about advanced event selectors, see Logging data events for trails in the CloudTrail User Guide.

trailName :: Text

Specifies the name of the trail or trail ARN. If you specify a trail name, the string must meet the following requirements:

Contain only ASCII letters (a-z, A-Z), numbers (0-9), periods (.), underscores (_), or dashes (-)

Start with a letter or number, and end with a letter or number

Be between 3 and 128 characters

Have no adjacent periods, underscores or dashes. Names like my-_namespace and my--namespace are not valid.

Not be in IP address format (for example, 192.168.5.4)

If you specify a trail ARN, it must be in the following format.

arn:aws:cloudtrail:us-east-2:123456789012:trail/MyTrail

newPutEventSelectors

Arguments

:: Text
$sel:trailName:PutEventSelectors'

-> PutEventSelectors

Create a value of PutEventSelectors with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventSelectors:PutEventSelectors', putEventSelectors_eventSelectors - Specifies the settings for your event selectors. You can configure up to five event selectors for a trail. You can use either EventSelectors or AdvancedEventSelectors in a PutEventSelectors request, but not both. If you apply EventSelectors to a trail, any existing AdvancedEventSelectors are overwritten.

$sel:advancedEventSelectors:PutEventSelectors', putEventSelectors_advancedEventSelectors - Specifies the settings for advanced event selectors. You can add advanced event selectors, and conditions for your advanced event selectors, up to a maximum of 500 values for all conditions and selectors on a trail. You can use either AdvancedEventSelectors or EventSelectors, but not both. If you apply AdvancedEventSelectors to a trail, any existing EventSelectors are overwritten. For more information about advanced event selectors, see Logging data events for trails in the CloudTrail User Guide.

$sel:trailName:PutEventSelectors', putEventSelectors_trailName - Specifies the name of the trail or trail ARN. If you specify a trail name, the string must meet the following requirements:

Contain only ASCII letters (a-z, A-Z), numbers (0-9), periods (.), underscores (_), or dashes (-)

Start with a letter or number, and end with a letter or number

Be between 3 and 128 characters

Have no adjacent periods, underscores or dashes. Names like my-_namespace and my--namespace are not valid.

Not be in IP address format (for example, 192.168.5.4)

If you specify a trail ARN, it must be in the following format.

arn:aws:cloudtrail:us-east-2:123456789012:trail/MyTrail

Request Lenses

putEventSelectors_eventSelectors :: Lens' PutEventSelectors (Maybe [EventSelector])

Specifies the settings for your event selectors. You can configure up to five event selectors for a trail. You can use either EventSelectors or AdvancedEventSelectors in a PutEventSelectors request, but not both. If you apply EventSelectors to a trail, any existing AdvancedEventSelectors are overwritten.

putEventSelectors_advancedEventSelectors :: Lens' PutEventSelectors (Maybe [AdvancedEventSelector])

Specifies the settings for advanced event selectors. You can add advanced event selectors, and conditions for your advanced event selectors, up to a maximum of 500 values for all conditions and selectors on a trail. You can use either AdvancedEventSelectors or EventSelectors, but not both. If you apply AdvancedEventSelectors to a trail, any existing EventSelectors are overwritten. For more information about advanced event selectors, see Logging data events for trails in the CloudTrail User Guide.

putEventSelectors_trailName :: Lens' PutEventSelectors Text

Specifies the name of the trail or trail ARN. If you specify a trail name, the string must meet the following requirements:

Contain only ASCII letters (a-z, A-Z), numbers (0-9), periods (.), underscores (_), or dashes (-)

Start with a letter or number, and end with a letter or number

Be between 3 and 128 characters

Have no adjacent periods, underscores or dashes. Names like my-_namespace and my--namespace are not valid.

Not be in IP address format (for example, 192.168.5.4)

If you specify a trail ARN, it must be in the following format.

arn:aws:cloudtrail:us-east-2:123456789012:trail/MyTrail

Destructuring the Response

data PutEventSelectorsResponse

See: newPutEventSelectorsResponse smart constructor.

Constructors

PutEventSelectorsResponse'

Fields

trailARN :: Maybe Text

Specifies the ARN of the trail that was updated with event selectors. The following is the format of a trail ARN.

arn:aws:cloudtrail:us-east-2:123456789012:trail/MyTrail

eventSelectors :: Maybe [EventSelector]

Specifies the event selectors configured for your trail.

advancedEventSelectors :: Maybe [AdvancedEventSelector]

Specifies the advanced event selectors configured for your trail.

httpStatus :: Int

The response's http status code.

newPutEventSelectorsResponse

Arguments

:: Int
$sel:httpStatus:PutEventSelectorsResponse'

-> PutEventSelectorsResponse

Create a value of PutEventSelectorsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:trailARN:PutEventSelectorsResponse', putEventSelectorsResponse_trailARN - Specifies the ARN of the trail that was updated with event selectors. The following is the format of a trail ARN.

arn:aws:cloudtrail:us-east-2:123456789012:trail/MyTrail

$sel:eventSelectors:PutEventSelectors', putEventSelectorsResponse_eventSelectors - Specifies the event selectors configured for your trail.

$sel:advancedEventSelectors:PutEventSelectors', putEventSelectorsResponse_advancedEventSelectors - Specifies the advanced event selectors configured for your trail.

$sel:httpStatus:PutEventSelectorsResponse', putEventSelectorsResponse_httpStatus - The response's http status code.

Response Lenses

putEventSelectorsResponse_trailARN :: Lens' PutEventSelectorsResponse (Maybe Text)

Specifies the ARN of the trail that was updated with event selectors. The following is the format of a trail ARN.

arn:aws:cloudtrail:us-east-2:123456789012:trail/MyTrail

putEventSelectorsResponse_eventSelectors :: Lens' PutEventSelectorsResponse (Maybe [EventSelector])

Specifies the event selectors configured for your trail.

putEventSelectorsResponse_advancedEventSelectors :: Lens' PutEventSelectorsResponse (Maybe [AdvancedEventSelector])

Specifies the advanced event selectors configured for your trail.

putEventSelectorsResponse_httpStatus :: Lens' PutEventSelectorsResponse Int

The response's http status code.