Network.AWS.CloudTrail.PutInsightSelectors

Description

Lets you enable Insights event logging by specifying the Insights selectors that you want to enable on an existing trail. You also use PutInsightSelectors to turn off Insights event logging, by passing an empty list of insight types. The valid Insights event type in this release is ApiCallRateInsight.

Creating a Request

data PutInsightSelectors

See: newPutInsightSelectors smart constructor.

Constructors

PutInsightSelectors'

Fields

trailName :: Text

The name of the CloudTrail trail for which you want to change or add Insights selectors.

insightSelectors :: [InsightSelector]

A JSON string that contains the Insights types that you want to log on a trail. The valid Insights type in this release is ApiCallRateInsight.

newPutInsightSelectors

Arguments

:: Text
$sel:trailName:PutInsightSelectors'

-> PutInsightSelectors

Create a value of PutInsightSelectors with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:trailName:PutInsightSelectors', putInsightSelectors_trailName - The name of the CloudTrail trail for which you want to change or add Insights selectors.

$sel:insightSelectors:PutInsightSelectors', putInsightSelectors_insightSelectors - A JSON string that contains the Insights types that you want to log on a trail. The valid Insights type in this release is ApiCallRateInsight.

Request Lenses

putInsightSelectors_trailName :: Lens' PutInsightSelectors Text

The name of the CloudTrail trail for which you want to change or add Insights selectors.

putInsightSelectors_insightSelectors :: Lens' PutInsightSelectors [InsightSelector]

A JSON string that contains the Insights types that you want to log on a trail. The valid Insights type in this release is ApiCallRateInsight.

Destructuring the Response

data PutInsightSelectorsResponse

See: newPutInsightSelectorsResponse smart constructor.

Constructors

PutInsightSelectorsResponse'

Fields

trailARN :: Maybe Text

The Amazon Resource Name (ARN) of a trail for which you want to change or add Insights selectors.

insightSelectors :: Maybe [InsightSelector]

A JSON string that contains the Insights event types that you want to log on a trail. The valid Insights type in this release is ApiCallRateInsight.

httpStatus :: Int

The response's http status code.

newPutInsightSelectorsResponse

Arguments

:: Int
$sel:httpStatus:PutInsightSelectorsResponse'

-> PutInsightSelectorsResponse

Create a value of PutInsightSelectorsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:trailARN:PutInsightSelectorsResponse', putInsightSelectorsResponse_trailARN - The Amazon Resource Name (ARN) of a trail for which you want to change or add Insights selectors.

$sel:insightSelectors:PutInsightSelectors', putInsightSelectorsResponse_insightSelectors - A JSON string that contains the Insights event types that you want to log on a trail. The valid Insights type in this release is ApiCallRateInsight.

$sel:httpStatus:PutInsightSelectorsResponse', putInsightSelectorsResponse_httpStatus - The response's http status code.

Response Lenses

putInsightSelectorsResponse_trailARN :: Lens' PutInsightSelectorsResponse (Maybe Text)

The Amazon Resource Name (ARN) of a trail for which you want to change or add Insights selectors.

putInsightSelectorsResponse_insightSelectors :: Lens' PutInsightSelectorsResponse (Maybe [InsightSelector])

A JSON string that contains the Insights event types that you want to log on a trail. The valid Insights type in this release is ApiCallRateInsight.

putInsightSelectorsResponse_httpStatus :: Lens' PutInsightSelectorsResponse Int

The response's http status code.