Network.AWS.CloudTrail.StopLogging

Description

Suspends the recording of Amazon Web Services API calls and log file delivery for the specified trail. Under most circumstances, there is no need to use this action. You can update a trail without stopping it first. This action is the only way to stop recording. For a trail enabled in all regions, this operation must be called from the region in which the trail was created, or an InvalidHomeRegionException will occur. This operation cannot be called on the shadow trails (replicated trails in other regions) of a trail enabled in all regions.

Creating a Request

data StopLogging

Passes the request to CloudTrail to stop logging Amazon Web Services API calls for the specified account.

See: newStopLogging smart constructor.

Constructors

StopLogging'

Fields

name :: Text

Specifies the name or the CloudTrail ARN of the trail for which CloudTrail will stop logging Amazon Web Services API calls. The following is the format of a trail ARN.

arn:aws:cloudtrail:us-east-2:123456789012:trail/MyTrail

newStopLogging

Arguments

:: Text
$sel:name:StopLogging'

-> StopLogging

Create a value of StopLogging with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:StopLogging', stopLogging_name - Specifies the name or the CloudTrail ARN of the trail for which CloudTrail will stop logging Amazon Web Services API calls. The following is the format of a trail ARN.

arn:aws:cloudtrail:us-east-2:123456789012:trail/MyTrail

Request Lenses

stopLogging_name :: Lens' StopLogging Text

Specifies the name or the CloudTrail ARN of the trail for which CloudTrail will stop logging Amazon Web Services API calls. The following is the format of a trail ARN.

arn:aws:cloudtrail:us-east-2:123456789012:trail/MyTrail

Destructuring the Response

data StopLoggingResponse

Returns the objects or data listed below if successful. Otherwise, returns an error.

See: newStopLoggingResponse smart constructor.

Constructors

StopLoggingResponse'

Fields

httpStatus :: Int

The response's http status code.

newStopLoggingResponse

Arguments

:: Int
$sel:httpStatus:StopLoggingResponse'

-> StopLoggingResponse

Create a value of StopLoggingResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:StopLoggingResponse', stopLoggingResponse_httpStatus - The response's http status code.

Response Lenses

stopLoggingResponse_httpStatus :: Lens' StopLoggingResponse Int

The response's http status code.