Network.AWS.CloudTrail.Types.AdvancedFieldSelector

Description

Documentation

data AdvancedFieldSelector

A single selector statement in an advanced event selector.

See: newAdvancedFieldSelector smart constructor.

Constructors

AdvancedFieldSelector'

Fields

notStartsWith :: Maybe (NonEmpty Text)

An operator that excludes events that match the first few characters of the event record field specified as the value of Field.

notEndsWith :: Maybe (NonEmpty Text)

An operator that excludes events that match the last few characters of the event record field specified as the value of Field.

notEquals :: Maybe (NonEmpty Text)

An operator that excludes events that match the exact value of the event record field specified as the value of Field.

equals :: Maybe (NonEmpty Text)

An operator that includes events that match the exact value of the event record field specified as the value of Field. This is the only valid operator that you can use with the readOnly, eventCategory, and resources.type fields.

startsWith :: Maybe (NonEmpty Text)

An operator that includes events that match the first few characters of the event record field specified as the value of Field.

endsWith :: Maybe (NonEmpty Text)

An operator that includes events that match the last few characters of the event record field specified as the value of Field.

field :: Text

A field in an event record on which to filter events to be logged. Supported fields include readOnly, eventCategory, eventSource (for management events), eventName, resources.type, and resources.ARN.

readOnly - Optional. Can be set to Equals a value of true or false. A value of false logs both read and write events.

eventSource - For filtering management events only. This can be set only to NotEquals kms.amazonaws.com.

eventName - Can use any operator. You can use it to ﬁlter in or ﬁlter out any data event logged to CloudTrail, such as PutBucket or GetSnapshotBlock. You can have multiple values for this ﬁeld, separated by commas.

eventCategory - This is required. It must be set to Equals, and the value must be Management or Data.

resources.type - This ﬁeld is required. resources.type can only use the Equals operator, and the value can be one of the following: AWS::S3::Object, AWS::S3::AccessPoint, AWS::Lambda::Function, AWS::DynamoDB::Table, AWS::S3Outposts::Object, AWS::ManagedBlockchain::Node, AWS::S3ObjectLambda::AccessPoint, or AWS::EC2::Snapshot. You can have only one resources.type ﬁeld per selector. To log data events on more than one resource type, add another selector.

resources.ARN - You can use any operator with resources.ARN, but if you use Equals or NotEquals, the value must exactly match the ARN of a valid resource of the type you've speciﬁed in the template as the value of resources.type. For example, if resources.type equals AWS::S3::Object, the ARN must be in one of the following formats. To log all data events for all objects in a specific S3 bucket, use the StartsWith operator, and include only the bucket ARN as the matching value.

The trailing slash is intentional; do not exclude it. Replace the text between less than and greater than symbols (<>) with resource-specific information.

arn:<partition>:s3:::<bucket_name>/

arn:<partition>:s3:::<bucket_name>/<object_path>/

When resources.type equals AWS::S3::AccessPoint, and the operator is set to Equals or NotEquals, the ARN must be in one of the following formats. To log events on all objects in an S3 access point, we recommend that you use only the access point ARN, don’t include the object path, and use the StartsWith or NotStartsWith operators.

arn:<partition>:s3:<region>:<account_ID>:accesspoint/<access_point_name>

arn:<partition>:s3:<region>:<account_ID>:accesspoint/<access_point_name>/object/<object_path>

When resources.type equals AWS::Lambda::Function, and the operator is set to Equals or NotEquals, the ARN must be in the following format:

arn:<partition>:lambda:<region>:<account_ID>:function:<function_name>

When resources.type equals AWS::DynamoDB::Table, and the operator is set to Equals or NotEquals, the ARN must be in the following format:

arn:<partition>:dynamodb:<region>:<account_ID>:table:<table_name>

When resources.type equals AWS::S3Outposts::Object, and the operator is set to Equals or NotEquals, the ARN must be in the following format:

arn:<partition>:s3-outposts:<region>:<account_ID>:<object_path>

When resources.type equals AWS::ManagedBlockchain::Node, and the operator is set to Equals or NotEquals, the ARN must be in the following format:

arn:<partition>:managedblockchain:<region>:<account_ID>:nodes/<node_ID>

When resources.type equals AWS::S3ObjectLambda::AccessPoint, and the operator is set to Equals or NotEquals, the ARN must be in the following format:

arn:<partition>:s3-object-lambda:<region>:<account_ID>:accesspoint/<access_point_name>

When resources.type equals AWS::EC2::Snapshot, and the operator is set to Equals or NotEquals, the ARN must be in the following format:

arn:<partition>:ec2:<region>::snapshot/<snapshot_ID>

newAdvancedFieldSelector

Arguments

:: Text
$sel:field:AdvancedFieldSelector'

-> AdvancedFieldSelector

Create a value of AdvancedFieldSelector with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:notStartsWith:AdvancedFieldSelector', advancedFieldSelector_notStartsWith - An operator that excludes events that match the first few characters of the event record field specified as the value of Field.

$sel:notEndsWith:AdvancedFieldSelector', advancedFieldSelector_notEndsWith - An operator that excludes events that match the last few characters of the event record field specified as the value of Field.

$sel:notEquals:AdvancedFieldSelector', advancedFieldSelector_notEquals - An operator that excludes events that match the exact value of the event record field specified as the value of Field.

$sel:equals:AdvancedFieldSelector', advancedFieldSelector_equals - An operator that includes events that match the exact value of the event record field specified as the value of Field. This is the only valid operator that you can use with the readOnly, eventCategory, and resources.type fields.

$sel:startsWith:AdvancedFieldSelector', advancedFieldSelector_startsWith - An operator that includes events that match the first few characters of the event record field specified as the value of Field.

$sel:endsWith:AdvancedFieldSelector', advancedFieldSelector_endsWith - An operator that includes events that match the last few characters of the event record field specified as the value of Field.

$sel:field:AdvancedFieldSelector', advancedFieldSelector_field - A field in an event record on which to filter events to be logged. Supported fields include readOnly, eventCategory, eventSource (for management events), eventName, resources.type, and resources.ARN.

readOnly - Optional. Can be set to Equals a value of true or false. A value of false logs both read and write events.

eventSource - For filtering management events only. This can be set only to NotEquals kms.amazonaws.com.

eventName - Can use any operator. You can use it to ﬁlter in or ﬁlter out any data event logged to CloudTrail, such as PutBucket or GetSnapshotBlock. You can have multiple values for this ﬁeld, separated by commas.

eventCategory - This is required. It must be set to Equals, and the value must be Management or Data.

resources.type - This ﬁeld is required. resources.type can only use the Equals operator, and the value can be one of the following: AWS::S3::Object, AWS::S3::AccessPoint, AWS::Lambda::Function, AWS::DynamoDB::Table, AWS::S3Outposts::Object, AWS::ManagedBlockchain::Node, AWS::S3ObjectLambda::AccessPoint, or AWS::EC2::Snapshot. You can have only one resources.type ﬁeld per selector. To log data events on more than one resource type, add another selector.

resources.ARN - You can use any operator with resources.ARN, but if you use Equals or NotEquals, the value must exactly match the ARN of a valid resource of the type you've speciﬁed in the template as the value of resources.type. For example, if resources.type equals AWS::S3::Object, the ARN must be in one of the following formats. To log all data events for all objects in a specific S3 bucket, use the StartsWith operator, and include only the bucket ARN as the matching value.

The trailing slash is intentional; do not exclude it. Replace the text between less than and greater than symbols (<>) with resource-specific information.

arn:<partition>:s3:::<bucket_name>/

arn:<partition>:s3:::<bucket_name>/<object_path>/

When resources.type equals AWS::S3::AccessPoint, and the operator is set to Equals or NotEquals, the ARN must be in one of the following formats. To log events on all objects in an S3 access point, we recommend that you use only the access point ARN, don’t include the object path, and use the StartsWith or NotStartsWith operators.

arn:<partition>:s3:<region>:<account_ID>:accesspoint/<access_point_name>

arn:<partition>:s3:<region>:<account_ID>:accesspoint/<access_point_name>/object/<object_path>

When resources.type equals AWS::Lambda::Function, and the operator is set to Equals or NotEquals, the ARN must be in the following format:

arn:<partition>:lambda:<region>:<account_ID>:function:<function_name>

When resources.type equals AWS::DynamoDB::Table, and the operator is set to Equals or NotEquals, the ARN must be in the following format:

arn:<partition>:dynamodb:<region>:<account_ID>:table:<table_name>

When resources.type equals AWS::S3Outposts::Object, and the operator is set to Equals or NotEquals, the ARN must be in the following format:

arn:<partition>:s3-outposts:<region>:<account_ID>:<object_path>

When resources.type equals AWS::ManagedBlockchain::Node, and the operator is set to Equals or NotEquals, the ARN must be in the following format:

arn:<partition>:managedblockchain:<region>:<account_ID>:nodes/<node_ID>

When resources.type equals AWS::S3ObjectLambda::AccessPoint, and the operator is set to Equals or NotEquals, the ARN must be in the following format:

arn:<partition>:s3-object-lambda:<region>:<account_ID>:accesspoint/<access_point_name>

When resources.type equals AWS::EC2::Snapshot, and the operator is set to Equals or NotEquals, the ARN must be in the following format:

arn:<partition>:ec2:<region>::snapshot/<snapshot_ID>

advancedFieldSelector_notStartsWith :: Lens' AdvancedFieldSelector (Maybe (NonEmpty Text))

An operator that excludes events that match the first few characters of the event record field specified as the value of Field.

advancedFieldSelector_notEndsWith :: Lens' AdvancedFieldSelector (Maybe (NonEmpty Text))

An operator that excludes events that match the last few characters of the event record field specified as the value of Field.

advancedFieldSelector_notEquals :: Lens' AdvancedFieldSelector (Maybe (NonEmpty Text))

An operator that excludes events that match the exact value of the event record field specified as the value of Field.

advancedFieldSelector_equals :: Lens' AdvancedFieldSelector (Maybe (NonEmpty Text))

An operator that includes events that match the exact value of the event record field specified as the value of Field. This is the only valid operator that you can use with the readOnly, eventCategory, and resources.type fields.

advancedFieldSelector_startsWith :: Lens' AdvancedFieldSelector (Maybe (NonEmpty Text))

An operator that includes events that match the first few characters of the event record field specified as the value of Field.

advancedFieldSelector_endsWith :: Lens' AdvancedFieldSelector (Maybe (NonEmpty Text))

An operator that includes events that match the last few characters of the event record field specified as the value of Field.

advancedFieldSelector_field :: Lens' AdvancedFieldSelector Text

A field in an event record on which to filter events to be logged. Supported fields include readOnly, eventCategory, eventSource (for management events), eventName, resources.type, and resources.ARN.

readOnly - Optional. Can be set to Equals a value of true or false. A value of false logs both read and write events.

eventSource - For filtering management events only. This can be set only to NotEquals kms.amazonaws.com.

eventName - Can use any operator. You can use it to ﬁlter in or ﬁlter out any data event logged to CloudTrail, such as PutBucket or GetSnapshotBlock. You can have multiple values for this ﬁeld, separated by commas.

eventCategory - This is required. It must be set to Equals, and the value must be Management or Data.

resources.type - This ﬁeld is required. resources.type can only use the Equals operator, and the value can be one of the following: AWS::S3::Object, AWS::S3::AccessPoint, AWS::Lambda::Function, AWS::DynamoDB::Table, AWS::S3Outposts::Object, AWS::ManagedBlockchain::Node, AWS::S3ObjectLambda::AccessPoint, or AWS::EC2::Snapshot. You can have only one resources.type ﬁeld per selector. To log data events on more than one resource type, add another selector.

resources.ARN - You can use any operator with resources.ARN, but if you use Equals or NotEquals, the value must exactly match the ARN of a valid resource of the type you've speciﬁed in the template as the value of resources.type. For example, if resources.type equals AWS::S3::Object, the ARN must be in one of the following formats. To log all data events for all objects in a specific S3 bucket, use the StartsWith operator, and include only the bucket ARN as the matching value.

The trailing slash is intentional; do not exclude it. Replace the text between less than and greater than symbols (<>) with resource-specific information.

arn:<partition>:s3:::<bucket_name>/

arn:<partition>:s3:::<bucket_name>/<object_path>/

When resources.type equals AWS::S3::AccessPoint, and the operator is set to Equals or NotEquals, the ARN must be in one of the following formats. To log events on all objects in an S3 access point, we recommend that you use only the access point ARN, don’t include the object path, and use the StartsWith or NotStartsWith operators.

arn:<partition>:s3:<region>:<account_ID>:accesspoint/<access_point_name>

arn:<partition>:s3:<region>:<account_ID>:accesspoint/<access_point_name>/object/<object_path>

When resources.type equals AWS::Lambda::Function, and the operator is set to Equals or NotEquals, the ARN must be in the following format:

arn:<partition>:lambda:<region>:<account_ID>:function:<function_name>

When resources.type equals AWS::DynamoDB::Table, and the operator is set to Equals or NotEquals, the ARN must be in the following format:

arn:<partition>:dynamodb:<region>:<account_ID>:table:<table_name>

When resources.type equals AWS::S3Outposts::Object, and the operator is set to Equals or NotEquals, the ARN must be in the following format:

arn:<partition>:s3-outposts:<region>:<account_ID>:<object_path>

When resources.type equals AWS::ManagedBlockchain::Node, and the operator is set to Equals or NotEquals, the ARN must be in the following format:

arn:<partition>:managedblockchain:<region>:<account_ID>:nodes/<node_ID>

When resources.type equals AWS::S3ObjectLambda::AccessPoint, and the operator is set to Equals or NotEquals, the ARN must be in the following format:

arn:<partition>:s3-object-lambda:<region>:<account_ID>:accesspoint/<access_point_name>

When resources.type equals AWS::EC2::Snapshot, and the operator is set to Equals or NotEquals, the ARN must be in the following format:

arn:<partition>:ec2:<region>::snapshot/<snapshot_ID>