Network.AWS.CloudTrail.Types.PublicKey

Description

Documentation

data PublicKey

Contains information about a returned public key.

See: newPublicKey smart constructor.

Constructors

PublicKey'

Fields

validityStartTime :: Maybe POSIX

The starting time of validity of the public key.

value :: Maybe Base64

The DER encoded public key value in PKCS#1 format.

validityEndTime :: Maybe POSIX

The ending time of validity of the public key.

fingerprint :: Maybe Text

The fingerprint of the public key.

newPublicKey :: PublicKey

Create a value of PublicKey with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:validityStartTime:PublicKey', publicKey_validityStartTime - The starting time of validity of the public key.

$sel:value:PublicKey', publicKey_value - The DER encoded public key value in PKCS#1 format.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:validityEndTime:PublicKey', publicKey_validityEndTime - The ending time of validity of the public key.

$sel:fingerprint:PublicKey', publicKey_fingerprint - The fingerprint of the public key.

publicKey_validityStartTime :: Lens' PublicKey (Maybe UTCTime)

The starting time of validity of the public key.

publicKey_value :: Lens' PublicKey (Maybe ByteString)

The DER encoded public key value in PKCS#1 format.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

publicKey_validityEndTime :: Lens' PublicKey (Maybe UTCTime)

The ending time of validity of the public key.

publicKey_fingerprint :: Lens' PublicKey (Maybe Text)

The fingerprint of the public key.