Network.AWS.CloudTrail.V2013_11_01.CreateTrail

Description

From the command line, use create-subscription. Creates a trail that specifies the settings for delivery of log data to an Amazon S3 bucket. Support for passing Trail as a parameter ends as early as February 25, 2014. The request and response examples in this topic show the use of parameters as well as a Trail object. Until Trail is removed, you can use either Trail or the parameter list.

Documentation

createTrail :: CreateTrail

Minimum specification for a CreateTrail request.

data CreateTrail

Constructors

CreateTrail

Fields

_ctrIncludeGlobalServiceEvents :: Maybe Bool

Specifies whether the trail is publishing events from global services such as IAM to the log files.

_ctrS3KeyPrefix :: Maybe Text

Specifies the Amazon S3 key prefix that precedes the name of the bucket you have designated for log file delivery.

_ctrSnsTopicName :: Maybe Text

Specifies the name of the Amazon SNS topic defined for notification of log file delivery.

_ctrName :: Maybe Text

Specifies the name of the trail.

_ctrS3BucketName :: Maybe Text

Specifies the name of the Amazon S3 bucket designated for publishing log files.

_ctrTrail :: Maybe Trail

Support for passing a Trail object in the CreateTrail or UpdateTrail actions will end as early as February 15, 2014. Instead of the Trail object and its members, use the parameters listed for these actions.

ctrTrail :: Lens' CreateTrail (Maybe Trail)

ctrSnsTopicName :: Lens' CreateTrail (Maybe Text)

ctrS3KeyPrefix :: Lens' CreateTrail (Maybe Text)

ctrS3BucketName :: Lens' CreateTrail (Maybe Text)

ctrName :: Lens' CreateTrail (Maybe Text)

ctrIncludeGlobalServiceEvents :: Lens' CreateTrail (Maybe Bool)

data CreateTrailResponse

Constructors

CreateTrailResponse

Fields

_ctsIncludeGlobalServiceEvents :: Maybe Bool

Specifies whether the trail is publishing events from global services such as IAM to the log files.

_ctsS3KeyPrefix :: Maybe Text

Specifies the Amazon S3 key prefix that precedes the name of the bucket you have designated for log file delivery.

_ctsSnsTopicName :: Maybe Text

Specifies the name of the Amazon SNS topic defined for notification of log file delivery.

_ctsName :: Maybe Text

Specifies the name of the trail.

_ctsS3BucketName :: Maybe Text

Specifies the name of the Amazon S3 bucket designated for publishing log files.

_ctsTrail :: Maybe Trail

Support for passing a Trail object in the CreateTrail or UpdateTrail actions will end as early as February 15, 2014. Instead of the Trail object and its members, use the parameters listed for these actions.

ctsTrail :: Lens' CreateTrailResponse (Maybe Trail)

ctsSnsTopicName :: Lens' CreateTrailResponse (Maybe Text)

ctsS3KeyPrefix :: Lens' CreateTrailResponse (Maybe Text)

ctsS3BucketName :: Lens' CreateTrailResponse (Maybe Text)

ctsName :: Lens' CreateTrailResponse (Maybe Text)

ctsIncludeGlobalServiceEvents :: Lens' CreateTrailResponse (Maybe Bool)