Network.AWS.CloudTrail.V2013_11_01.UpdateTrail

Description

From the command line, use update-subscription. Updates the settings that specify delivery of log files. Changes to a trail do not require stopping the CloudTrail service. Use this action to designate an existing bucket for log delivery. If the existing bucket has previously been a target for CloudTrail log files, an IAM policy exists for the bucket. Support for passing Trail as a parameter ends as early as February 25, 2014. The request and response examples in this topic show the use of parameters as well as a Trail object. Until Trail is removed, you can use either Trail or the parameter list.

Documentation

updateTrail :: UpdateTrail

Minimum specification for a UpdateTrail request.

data UpdateTrail

Constructors

UpdateTrail

Fields

_utrIncludeGlobalServiceEvents :: Maybe Bool

Specifies whether the trail is publishing events from global services such as IAM to the log files.

_utrS3KeyPrefix :: Maybe Text

Specifies the Amazon S3 key prefix that precedes the name of the bucket you have designated for log file delivery.

_utrSnsTopicName :: Maybe Text

Specifies the name of the Amazon SNS topic defined for notification of log file delivery.

_utrName :: Maybe Text

Specifies the name of the trail.

_utrS3BucketName :: Maybe Text

Specifies the name of the Amazon S3 bucket designated for publishing log files.

_utrTrail :: Maybe Trail

Support for passing a Trail object in the CreateTrail or UpdateTrail actions will end as early as February 15, 2014. Instead of the Trail object and its members, use the parameters listed for these actions.

utrTrail :: Lens' UpdateTrail (Maybe Trail)

utrSnsTopicName :: Lens' UpdateTrail (Maybe Text)

utrS3KeyPrefix :: Lens' UpdateTrail (Maybe Text)

utrS3BucketName :: Lens' UpdateTrail (Maybe Text)

utrName :: Lens' UpdateTrail (Maybe Text)

utrIncludeGlobalServiceEvents :: Lens' UpdateTrail (Maybe Bool)

data UpdateTrailResponse

Constructors

UpdateTrailResponse

Fields

_utsIncludeGlobalServiceEvents :: Maybe Bool

Specifies whether the trail is publishing events from global services such as IAM to the log files.

_utsS3KeyPrefix :: Maybe Text

Specifies the Amazon S3 key prefix that precedes the name of the bucket you have designated for log file delivery.

_utsSnsTopicName :: Maybe Text

Specifies the name of the Amazon SNS topic defined for notification of log file delivery.

_utsName :: Maybe Text

Specifies the name of the trail.

_utsS3BucketName :: Maybe Text

Specifies the name of the Amazon S3 bucket designated for publishing log files.

_utsTrail :: Maybe Trail

Represents the CloudTrail settings that were updated by calling UpdateTrail.

utsTrail :: Lens' UpdateTrailResponse (Maybe Trail)

utsSnsTopicName :: Lens' UpdateTrailResponse (Maybe Text)

utsS3KeyPrefix :: Lens' UpdateTrailResponse (Maybe Text)

utsS3BucketName :: Lens' UpdateTrailResponse (Maybe Text)

utsName :: Lens' UpdateTrailResponse (Maybe Text)

utsIncludeGlobalServiceEvents :: Lens' UpdateTrailResponse (Maybe Bool)