Network.AWS.CloudWatch

Description

Amazon CloudWatch is a web service that enables you to publish, monitor, and manage various metrics, as well as configure alarm actions based on data from metrics.

Actions

DeleteAlarms

data DeleteAlarms

Deletes all specified alarms. In the event of an error, no alarms are deleted.

http://docs.aws.amazon.com/AmazonCloudWatch/latest/APIReference/API_DeleteAlarms.html

Constructors

DeleteAlarms

Fields

daAlarmNames :: Members Text

A list of alarms to be deleted.

data DeleteAlarmsResponse

Constructors

DeleteAlarmsResponse

Fields

darResponseMetadata :: !Text

DescribeAlarmHistory

data DescribeAlarmHistory

Retrieves history for the specified alarm. Filter alarms by date range or item type. If an alarm name is not specified, Amazon CloudWatch returns histories for all of the owner's alarms. Note Amazon CloudWatch retains the history of an alarm for two weeks, whether or not you delete the alarm.

http://docs.aws.amazon.com/AmazonCloudWatch/latest/APIReference/API_DescribeAlarmHistory.html

Constructors

DescribeAlarmHistory

Fields

dahAlarmName :: Maybe Text

The name of the alarm.

dahEndDate :: Maybe UTCTime

The ending date to retrieve alarm history.

dahHistoryItemType :: Maybe Text

The type of alarm histories to retrieve.

dahMaxRecords :: Maybe Integer

The maximum number of alarm history records to retrieve.

dahNextToken :: Maybe Text

The token returned by a previous call to indicate that there is more data available.

dahStartDate :: Maybe UTCTime

The starting date to retrieve alarm history.

data DescribeAlarmHistoryResponse

Constructors

DescribeAlarmHistoryResponse

Fields

dahrResponseMetadata :: !Text

dahrDescribeAlarmHistoryResult :: !DescribeAlarmHistoryResult

DescribeAlarms

data DescribeAlarms

Retrieves alarms with the specified names. If no name is specified, all alarms for the user are returned. Alarms can be retrieved by using only a prefix for the alarm name, the alarm state, or a prefix for any action.

http://docs.aws.amazon.com/AmazonCloudWatch/latest/APIReference/API_DescribeAlarms.html

Constructors

DescribeAlarms

Fields

daActionPrefix :: Maybe Text

The action name prefix.

daAlarmNamePrefix :: Maybe Text

The alarm name prefix. AlarmNames cannot be specified if this parameter is specified.

dbAlarmNames :: Members Text

A list of alarm names to retrieve information for.

dbMaxRecords :: Maybe Integer

The maximum number of alarm descriptions to retrieve.

dbNextToken :: Maybe Text

The token returned by a previous call to indicate that there is more data available.

dbStateValue :: Maybe Text

The state value to be used in matching alarms.

data DescribeAlarmsResponse

Constructors

DescribeAlarmsResponse

Fields

dasResponseMetadata :: !Text

dasDescribeAlarmsResult :: !DescribeAlarmsResult

DescribeAlarmsForMetric

data DescribeAlarmsForMetric

Retrieves all alarms for a single metric. Specify a statistic, period, or unit to filter the set of alarms further.

http://docs.aws.amazon.com/AmazonCloudWatch/latest/APIReference/API_DescribeAlarmsForMetric.html

Constructors

DescribeAlarmsForMetric

Fields

dafmDimensions :: Members Dimension

The list of dimensions associated with the metric.

dafmMetricName :: !Text

The name of the metric.

dafmNamespace :: !Text

The namespace of the metric.

dafmPeriod :: Maybe Integer

The period in seconds over which the statistic is applied.

dafmStatistic :: Maybe Text

The statistic for the metric.

dafmUnit :: Maybe Text

The unit for the metric.

data DescribeAlarmsForMetricResponse

Constructors

DescribeAlarmsForMetricResponse

Fields

dafmrResponseMetadata :: !Text

dafmrDescribeAlarmsForMetricResult :: !DescribeAlarmsForMetricResult

DisableAlarmActions

data DisableAlarmActions

Disables actions for the specified alarms. When an alarm's actions are disabled the alarm's state may change, but none of the alarm's actions will execute.

http://docs.aws.amazon.com/AmazonCloudWatch/latest/APIReference/API_DisableAlarmActions.html

Constructors

DisableAlarmActions

Fields

daaAlarmNames :: Members Text

The names of the alarms to disable actions for.

data DisableAlarmActionsResponse

Constructors

DisableAlarmActionsResponse

Fields

daarResponseMetadata :: !Text

EnableAlarmActions

data EnableAlarmActions

Enables actions for the specified alarms.

http://docs.aws.amazon.com/AmazonCloudWatch/latest/APIReference/API_EnableAlarmActions.html

Constructors

EnableAlarmActions

Fields

eaaAlarmNames :: Members Text

The names of the alarms to enable actions for.

data EnableAlarmActionsResponse

Constructors

EnableAlarmActionsResponse

Fields

eaarResponseMetadata :: !Text

GetMetricStatistics

data GetMetricStatistics

Gets statistics for the specified metric. The maximum number of data points returned from a single GetMetricStatistics request is 1,440, whereas the maximum number of data points that can be queried is 50,850. If you make a request that generates more than 1,440 data points, Amazon CloudWatch returns an error. In such a case, you can alter the request by narrowing the specified time range or increasing the specified period. Alternatively, you can make multiple requests across adjacent time ranges. GetMetricStatistics does not return the data in chronological order. Amazon CloudWatch aggregates data points based on the length of the period that you specify. For example, if you request statistics with a one-minute granularity, Amazon CloudWatch aggregates data points with time stamps that fall within the same one-minute period. In such a case, the data points queried can greatly outnumber the data points returned. The following examples show various statistics allowed by the data point query maximum of 50,850 when you call GetMetricStatistics on Amazon EC2 instances with detailed (one-minute) monitoring enabled: For information about the namespace, metric names, and dimensions that other Amazon Web Services products use to send metrics to CloudWatch, go to Amazon CloudWatch Metrics, Namespaces, and Dimensions Reference in the Amazon CloudWatch Developer Guide.

http://docs.aws.amazon.com/AmazonCloudWatch/latest/APIReference/API_GetMetricStatistics.html

Constructors

GetMetricStatistics

Fields

gmsDimensions :: Members Dimension

A list of dimensions describing qualities of the metric.

gmsEndTime :: !UTCTime

The time stamp to use for determining the last datapoint to return. The value specified is exclusive; results will include datapoints up to the time stamp specified.

gmsMetricName :: !Text

The name of the metric, with or without spaces.

gmsNamespace :: !Text

The namespace of the metric, with or without spaces.

gmsPeriod :: !Integer

The granularity, in seconds, of the returned datapoints. Period must be at least 60 seconds and must be a multiple of 60. The default value is 60.

gmsStartTime :: !UTCTime

The time stamp to use for determining the first datapoint to return. The value specified is inclusive; results include datapoints with the time stamp specified.

gmsStatistics :: Members Text

The metric statistics to return. For information about specific statistics returned by GetMetricStatistics, go to Statistics in the Amazon CloudWatch Developer Guide.

gmsUnit :: Maybe Text

The unit for the metric.

data GetMetricStatisticsResponse

Constructors

GetMetricStatisticsResponse

Fields

gmsrResponseMetadata :: !Text

gmsrGetMetricStatisticsResult :: !GetMetricStatisticsResult

ListMetrics

data ListMetrics

Returns a list of valid metrics stored for the AWS account owner. Returned metrics can be used with GetMetricStatistics to obtain statistical data for a given metric. Note Up to 500 results are returned for any one call. To retrieve further results, use returned NextToken values with subsequent ListMetrics operations. Note If you create a metric with the PutMetricData action, allow up to fifteen minutes for the metric to appear in calls to the ListMetrics action. Statistics about the metric, however, are available sooner using GetMetricStatistics.

http://docs.aws.amazon.com/AmazonCloudWatch/latest/APIReference/API_ListMetrics.html

Constructors

ListMetrics

Fields

lmDimensions :: Members DimensionFilter

A list of dimensions to filter against.

lmMetricName :: Maybe Text

The name of the metric to filter against.

lmNamespace :: Maybe Text

The namespace to filter against.

lmNextToken :: Maybe Text

The token returned by a previous call to indicate that there is more data available.

data ListMetricsResponse

Constructors

ListMetricsResponse

Fields

lmrResponseMetadata :: !Text

lmrListMetricsResult :: !ListMetricsResult

PutMetricAlarm

data PutMetricAlarm

Creates or updates an alarm and associates it with the specified Amazon CloudWatch metric. Optionally, this operation can associate one or more Amazon Simple Notification Service resources with the alarm. When this operation creates an alarm, the alarm state is immediately set to INSUFFICIENT_DATA. The alarm is evaluated and its StateValue is set appropriately. Any actions associated with the StateValue is then executed. Note When updating an existing alarm, its StateValue is left unchanged.

http://docs.aws.amazon.com/AmazonCloudWatch/latest/APIReference/API_PutMetricAlarm.html

Constructors

PutMetricAlarm

Fields

pmaActionsEnabled :: Maybe Bool

Indicates whether or not actions should be executed during any changes to the alarm's state.

pmaAlarmActions :: Members Text

The list of actions to execute when this alarm transitions into an ALARM state from any other state. Each action is specified as an Amazon Resource Number (ARN). Currently the only action supported is publishing to an Amazon SNS topic or an Amazon Auto Scaling policy.

pmaAlarmDescription :: Maybe Text

The description for the alarm.

pmaAlarmName :: !Text

The descriptive name for the alarm. This name must be unique within the user's AWS account

pmaComparisonOperator :: !Text

The arithmetic operation to use when comparing the specified Statistic and Threshold. The specified Statistic value is used as the first operand.

pmaDimensions :: Members Dimension

The dimensions for the alarm's associated metric.

pmaEvaluationPeriods :: !Integer

The number of periods over which data is compared to the specified threshold.

pmaInsufficientDataActions :: Members Text

The list of actions to execute when this alarm transitions into an INSUFFICIENT_DATA state from any other state. Each action is specified as an Amazon Resource Number (ARN). Currently the only action supported is publishing to an Amazon SNS topic or an Amazon Auto Scaling policy.

pmaMetricName :: !Text

The name for the alarm's associated metric.

pmaNamespace :: !Text

The namespace for the alarm's associated metric.

pmaOKActions :: Members Text

The list of actions to execute when this alarm transitions into an OK state from any other state. Each action is specified as an Amazon Resource Number (ARN). Currently the only action supported is publishing to an Amazon SNS topic or an Amazon Auto Scaling policy.

pmaPeriod :: !Integer

The period in seconds over which the specified statistic is applied.

pmaStatistic :: !Text

The statistic to apply to the alarm's associated metric.

pmaThreshold :: !Double

The value against which the specified statistic is compared.

pmaUnit :: Maybe Text

The unit for the alarm's associated metric.

data PutMetricAlarmResponse

Constructors

PutMetricAlarmResponse

Fields

pmarResponseMetadata :: !Text

PutMetricData

data PutMetricData

Publishes metric data points to Amazon CloudWatch. Amazon CloudWatch associates the data points with the specified metric. If the specified metric does not exist, Amazon CloudWatch creates the metric. When Amazon CloudWatch creates a metric, it can take up to fifteen minutes for the metric to appear in calls to the ListMetrics action. Each PutMetricData request is limited to 8 KB in size for HTTP GET requests and is limited to 40 KB in size for HTTP POST requests. Important Although the Value parameter accepts numbers of type Double, Amazon CloudWatch rejects values that are either too small or too large. Values must be in the range of 8.515920e-109 to 1.174271e+108 (Base 10) or 2e-360 to 2e360 (Base 2). In addition, special values (e.g., NaN, +Infinity, -Infinity) are not supported. Data that is timestamped 24 hours or more in the past may take in excess of 48 hours to become available from submission time using GetMetricStatistics.

http://docs.aws.amazon.com/AmazonCloudWatch/latest/APIReference/API_PutMetricData.html

Constructors

PutMetricData

Fields

pmdMetricData :: Members MetricDatum

A list of data describing the metric.

pmdNamespace :: !Text

The namespace for the metric data.

data PutMetricDataResponse

Constructors

PutMetricDataResponse

Fields

pmdrResponseMetadata :: !Text

SetAlarmState

data SetAlarmState

Temporarily sets the state of an alarm. When the updated StateValue differs from the previous value, the action configured for the appropriate state is invoked. This is not a permanent change. The next periodic alarm check (in about a minute) will set the alarm to its actual state.

http://docs.aws.amazon.com/AmazonCloudWatch/latest/APIReference/API_SetAlarmState.html

Constructors

SetAlarmState

Fields

sasAlarmName :: !Text

The descriptive name for the alarm. This name must be unique within the user's AWS account. The maximum length is 255 characters.

sasStateReason :: !Text

The reason that this alarm is set to this specific state (in human-readable text format)

sasStateReasonData :: Maybe Text

The reason that this alarm is set to this specific state (in machine-readable JSON format)

sasStateValue :: !Text

The value of the state.

data SetAlarmStateResponse

Constructors

SetAlarmStateResponse

Fields

sasrResponseMetadata :: !Text

Data Types

cloudWatch :: Service

Currently supported version of the CloudWatch service.

data CloudWatchError

Constructors

CloudWatchError

Fields

cweCode :: !Text

cweMessage :: !Text

data AlarmHistoryItem

The AlarmHistoryItem data type contains descriptive information about the history of a specific alarm. If you call DescribeAlarmHistory, Amazon CloudWatch returns this data type as part of the DescribeAlarmHistoryResult data type.

http://docs.aws.amazon.com/AmazonCloudWatch/latest/APIReference/API_AlarmHistoryItem.html

Constructors

AlarmHistoryItem

Fields

ahiAlarmName :: Maybe Text

The descriptive name for the alarm.

ahiHistoryData :: Maybe Text

Machine-readable data about the alarm in JSON format.

ahiHistoryItemType :: Maybe Text

The type of alarm history item.

ahiHistorySummary :: Maybe Text

A human-readable summary of the alarm history.

ahiTimestamp :: Maybe UTCTime

The time stamp for the alarm history item.

data Datapoint

The Datapoint data type encapsulates the statistical data that Amazon CloudWatch computes from metric data.

http://docs.aws.amazon.com/AmazonCloudWatch/latest/APIReference/API_Datapoint.html

Constructors

Datapoint

Fields

dAverage :: Maybe Double

The average of metric values that correspond to the datapoint.

dMaximum :: Maybe Double

The maximum of the metric value used for the datapoint.

dMinimum :: Maybe Double

The minimum metric value used for the datapoint.

dSampleCount :: Maybe Double

The number of metric values that contributed to the aggregate value of this datapoint.

dSum :: Maybe Double

The sum of metric values used for the datapoint.

dTimestamp :: Maybe UTCTime

The time stamp used for the datapoint.

dUnit :: Maybe Text

The standard unit used for the datapoint.

data DescribeAlarmHistoryResult

The output for the DescribeAlarmHistory action.

http://docs.aws.amazon.com/AmazonCloudWatch/latest/APIReference/API_DescribeAlarmHistoryResult.html

Constructors

DescribeAlarmHistoryResult

Fields

dahrAlarmHistoryItems :: Maybe AlarmHistoryItem

A list of alarm histories in JSON format.

dahrNextToken :: Maybe Text

A string that marks the start of the next batch of returned results.

data DescribeAlarmsForMetricResult

The output for the DescribeAlarmsForMetric action.

http://docs.aws.amazon.com/AmazonCloudWatch/latest/APIReference/API_DescribeAlarmsForMetricResult.html

Constructors

DescribeAlarmsForMetricResult

Fields

dafmrMetricAlarms :: Maybe MetricAlarm

A list of information for each alarm with the specified metric.

data DescribeAlarmsResult

The output for the DescribeAlarms action.

http://docs.aws.amazon.com/AmazonCloudWatch/latest/APIReference/API_DescribeAlarmsResult.html

Constructors

DescribeAlarmsResult

Fields

darMetricAlarms :: Maybe MetricAlarm

A list of information for the specified alarms.

darNextToken :: Maybe Text

A string that marks the start of the next batch of returned results.

data Dimension

The Dimension data type further expands on the identity of a metric using a Name, Value pair. For examples that use one or more dimensions, see PutMetricData.

http://docs.aws.amazon.com/AmazonCloudWatch/latest/APIReference/API_Dimension.html

Constructors

Dimension

Fields

dName :: !Text

The name of the dimension.

dValue :: !Text

The value representing the dimension measurement

data DimensionFilter

The DimensionFilter data type is used to filter ListMetrics results.

http://docs.aws.amazon.com/AmazonCloudWatch/latest/APIReference/API_DimensionFilter.html

Constructors

DimensionFilter

Fields

dfName :: !Text

The dimension name to be matched.

dfValue :: Maybe Text

The value of the dimension to be matched.

data GetMetricStatisticsResult

The output for the GetMetricStatistics action.

http://docs.aws.amazon.com/AmazonCloudWatch/latest/APIReference/API_GetMetricStatisticsResult.html

Constructors

GetMetricStatisticsResult

Fields

gmsrDatapoints :: Maybe Datapoint

The datapoints for the specified metric.

gmsrLabel :: Maybe Text

A label describing the specified metric.

data ListMetricsResult

The output for the ListMetrics action.

http://docs.aws.amazon.com/AmazonCloudWatch/latest/APIReference/API_ListMetricsResult.html

Constructors

ListMetricsResult

Fields

lmrMetrics :: Maybe Metric

A list of metrics used to generate statistics for an AWS account.

lmrNextToken :: Maybe Text

A string that marks the start of the next batch of returned results.

data Metric

The Metric data type contains information about a specific metric. If you call ListMetrics, Amazon CloudWatch returns information contained by this data type. The example in the Examples section publishes two metrics named buffers and latency. Both metrics are in the examples namespace. Both metrics have two dimensions, InstanceID and InstanceType.

http://docs.aws.amazon.com/AmazonCloudWatch/latest/APIReference/API_Metric.html

Constructors

Metric

Fields

mDimensions :: Maybe Dimension

A list of dimensions associated with the metric.

mMetricName :: Maybe Text

The name of the metric.

mNamespace :: Maybe Text

The namespace of the metric.

data MetricAlarm

The MetricAlarm data type represents an alarm. You can use PutMetricAlarm to create or update an alarm.

http://docs.aws.amazon.com/AmazonCloudWatch/latest/APIReference/API_MetricAlarm.html

Constructors

MetricAlarm

Fields

maActionsEnabled :: Maybe Bool

Indicates whether actions should be executed during any changes to the alarm's state.

maAlarmActions :: Maybe Text

The list of actions to execute when this alarm transitions into an ALARM state from any other state. Each action is specified as an Amazon Resource Number (ARN). Currently the only actions supported are publishing to an Amazon SNS topic and triggering an Auto Scaling policy.

maAlarmArn :: Maybe Text

The Amazon Resource Name (ARN) of the alarm.

maAlarmConfigurationUpdatedTimestamp :: Maybe UTCTime

The time stamp of the last update to the alarm configuration.

maAlarmDescription :: Maybe Text

The description for the alarm.

maAlarmName :: Maybe Text

The name of the alarm.

maComparisonOperator :: Maybe Text

The arithmetic operation to use when comparing the specified Statistic and Threshold. The specified Statistic value is used as the first operand.

maDimensions :: Maybe Dimension

The list of dimensions associated with the alarm's associated metric.

maEvaluationPeriods :: Maybe Integer

The number of periods over which data is compared to the specified threshold.

maInsufficientDataActions :: Maybe Text

The list of actions to execute when this alarm transitions into an INSUFFICIENT_DATA state from any other state. Each action is specified as an Amazon Resource Number (ARN). Currently the only actions supported are publishing to an Amazon SNS topic or triggering an Auto Scaling policy.

maMetricName :: Maybe Text

The name of the alarm's metric.

maNamespace :: Maybe Text

The namespace of alarm's associated metric.

maOKActions :: Maybe Text

The list of actions to execute when this alarm transitions into an OK state from any other state. Each action is specified as an Amazon Resource Number (ARN). Currently the only actions supported are publishing to an Amazon SNS topic and triggering an Auto Scaling policy.

maPeriod :: Maybe Integer

The period in seconds over which the statistic is applied.

maStateReason :: Maybe Text

A human-readable explanation for the alarm's state.

maStateReasonData :: Maybe Text

An explanation for the alarm's state in machine-readable JSON format

maStateUpdatedTimestamp :: Maybe UTCTime

The time stamp of the last update to the alarm's state.

maStateValue :: Maybe Text

The state value for the alarm.

maStatistic :: Maybe Text

The statistic to apply to the alarm's associated metric.

maThreshold :: Maybe Double

The value against which the specified statistic is compared.

maUnit :: Maybe Text

The unit of the alarm's associated metric.

data MetricDatum

The MetricDatum data type encapsulates the information sent with PutMetricData to either create a new metric or add new values to be aggregated into an existing metric.

http://docs.aws.amazon.com/AmazonCloudWatch/latest/APIReference/API_MetricDatum.html

Constructors

MetricDatum

Fields

mdDimensions :: Maybe Dimension

A list of dimensions associated with the metric. Note, when using the Dimensions value in a query, you need to append .member.N to it (e.g., Dimensions.member.N).

mdMetricName :: !Text

The name of the metric.

mdStatisticValues :: Maybe StatisticSet

A set of statistical values describing the metric.

mdTimestamp :: Maybe UTCTime

The time stamp used for the metric in ISO 8601 Universal Coordinated Time (UTC) format. If not specified, the default value is set to the time the metric data was received.

mdUnit :: Maybe Text

The unit of the metric.

mdValue :: Maybe Double

The value for the metric.

data StatisticSet

The StatisticSet data type describes the StatisticValues component of MetricDatum, and represents a set of statistics that describes a specific metric.

http://docs.aws.amazon.com/AmazonCloudWatch/latest/APIReference/API_StatisticSet.html

Constructors

StatisticSet

Fields

ssMaximum :: !Double

The maximum value of the sample set.

ssMinimum :: !Double

The minimum value of the sample set.

ssSampleCount :: !Double

The number of samples used for the statistic set.

ssSum :: !Double

The sum of values for the sample set.

cloudWatchNS :: ByteString

XML namespace to annotate CloudWatch elements with.

cloudWatchElem :: ByteString -> NName ByteString

Helper to define CloudWatch namespaced XML elements.

Common

module Network.AWS