Network.AWS.CloudWatch.DeleteAlarms

Description

Deletes the specified alarms. You can delete up to 100 alarms in one operation. However, this total can include no more than one composite alarm. For example, you could delete 99 metric alarms and one composite alarms with one operation, but you can't delete two composite alarms with one operation.

In the event of an error, no alarms are deleted.

It is possible to create a loop or cycle of composite alarms, where composite alarm A depends on composite alarm B, and composite alarm B also depends on composite alarm A. In this scenario, you can't delete any composite alarm that is part of the cycle because there is always still a composite alarm that depends on that alarm that you want to delete.

To get out of such a situation, you must break the cycle by changing the rule of one of the composite alarms in the cycle to remove a dependency that creates the cycle. The simplest change to make to break a cycle is to change the AlarmRule of one of the alarms to False.

Additionally, the evaluation of composite alarms stops if CloudWatch detects a cycle in the evaluation path.

Creating a Request

data DeleteAlarms

See: newDeleteAlarms smart constructor.

Constructors

DeleteAlarms'

Fields

alarmNames :: [Text]

The alarms to be deleted.

newDeleteAlarms :: DeleteAlarms

Create a value of DeleteAlarms with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:alarmNames:DeleteAlarms', deleteAlarms_alarmNames - The alarms to be deleted.

Request Lenses

deleteAlarms_alarmNames :: Lens' DeleteAlarms [Text]

The alarms to be deleted.

Destructuring the Response

data DeleteAlarmsResponse

See: newDeleteAlarmsResponse smart constructor.

Constructors

DeleteAlarmsResponse'

newDeleteAlarmsResponse :: DeleteAlarmsResponse

Create a value of DeleteAlarmsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.