Network.AWS.CloudWatch.DescribeAlarmHistory

Description

Retrieves the history for the specified alarm. You can filter the results by date range or item type. If an alarm name is not specified, the histories for either all metric alarms or all composite alarms are returned.

CloudWatch retains the history of an alarm even if you delete the alarm.

This operation returns paginated results.

Creating a Request

data DescribeAlarmHistory

See: newDescribeAlarmHistory smart constructor.

Constructors

DescribeAlarmHistory'

Fields

alarmName :: Maybe Text

The name of the alarm.

historyItemType :: Maybe HistoryItemType

The type of alarm histories to retrieve.

alarmTypes :: Maybe [AlarmType]

Use this parameter to specify whether you want the operation to return metric alarms or composite alarms. If you omit this parameter, only metric alarms are returned.

endDate :: Maybe ISO8601

The ending date to retrieve alarm history.

startDate :: Maybe ISO8601

The starting date to retrieve alarm history.

nextToken :: Maybe Text

The token returned by a previous call to indicate that there is more data available.

scanBy :: Maybe ScanBy

Specified whether to return the newest or oldest alarm history first. Specify TimestampDescending to have the newest event history returned first, and specify TimestampAscending to have the oldest history returned first.

maxRecords :: Maybe Natural

The maximum number of alarm history records to retrieve.

newDescribeAlarmHistory :: DescribeAlarmHistory

Create a value of DescribeAlarmHistory with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:alarmName:DescribeAlarmHistory', describeAlarmHistory_alarmName - The name of the alarm.

$sel:historyItemType:DescribeAlarmHistory', describeAlarmHistory_historyItemType - The type of alarm histories to retrieve.

$sel:alarmTypes:DescribeAlarmHistory', describeAlarmHistory_alarmTypes - Use this parameter to specify whether you want the operation to return metric alarms or composite alarms. If you omit this parameter, only metric alarms are returned.

$sel:endDate:DescribeAlarmHistory', describeAlarmHistory_endDate - The ending date to retrieve alarm history.

$sel:startDate:DescribeAlarmHistory', describeAlarmHistory_startDate - The starting date to retrieve alarm history.

$sel:nextToken:DescribeAlarmHistory', describeAlarmHistory_nextToken - The token returned by a previous call to indicate that there is more data available.

$sel:scanBy:DescribeAlarmHistory', describeAlarmHistory_scanBy - Specified whether to return the newest or oldest alarm history first. Specify TimestampDescending to have the newest event history returned first, and specify TimestampAscending to have the oldest history returned first.

$sel:maxRecords:DescribeAlarmHistory', describeAlarmHistory_maxRecords - The maximum number of alarm history records to retrieve.

Request Lenses

describeAlarmHistory_alarmName :: Lens' DescribeAlarmHistory (Maybe Text)

The name of the alarm.

describeAlarmHistory_historyItemType :: Lens' DescribeAlarmHistory (Maybe HistoryItemType)

The type of alarm histories to retrieve.

describeAlarmHistory_alarmTypes :: Lens' DescribeAlarmHistory (Maybe [AlarmType])

Use this parameter to specify whether you want the operation to return metric alarms or composite alarms. If you omit this parameter, only metric alarms are returned.

describeAlarmHistory_endDate :: Lens' DescribeAlarmHistory (Maybe UTCTime)

The ending date to retrieve alarm history.

describeAlarmHistory_startDate :: Lens' DescribeAlarmHistory (Maybe UTCTime)

The starting date to retrieve alarm history.

describeAlarmHistory_nextToken :: Lens' DescribeAlarmHistory (Maybe Text)

The token returned by a previous call to indicate that there is more data available.

describeAlarmHistory_scanBy :: Lens' DescribeAlarmHistory (Maybe ScanBy)

Specified whether to return the newest or oldest alarm history first. Specify TimestampDescending to have the newest event history returned first, and specify TimestampAscending to have the oldest history returned first.

describeAlarmHistory_maxRecords :: Lens' DescribeAlarmHistory (Maybe Natural)

The maximum number of alarm history records to retrieve.

Destructuring the Response

data DescribeAlarmHistoryResponse

See: newDescribeAlarmHistoryResponse smart constructor.

Constructors

DescribeAlarmHistoryResponse'

Fields

alarmHistoryItems :: Maybe [AlarmHistoryItem]

The alarm histories, in JSON format.

nextToken :: Maybe Text

The token that marks the start of the next batch of returned results.

httpStatus :: Int

The response's http status code.

newDescribeAlarmHistoryResponse

Arguments

:: Int
$sel:httpStatus:DescribeAlarmHistoryResponse'

-> DescribeAlarmHistoryResponse

Create a value of DescribeAlarmHistoryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:alarmHistoryItems:DescribeAlarmHistoryResponse', describeAlarmHistoryResponse_alarmHistoryItems - The alarm histories, in JSON format.

$sel:nextToken:DescribeAlarmHistory', describeAlarmHistoryResponse_nextToken - The token that marks the start of the next batch of returned results.

$sel:httpStatus:DescribeAlarmHistoryResponse', describeAlarmHistoryResponse_httpStatus - The response's http status code.

Response Lenses

describeAlarmHistoryResponse_alarmHistoryItems :: Lens' DescribeAlarmHistoryResponse (Maybe [AlarmHistoryItem])

The alarm histories, in JSON format.

describeAlarmHistoryResponse_nextToken :: Lens' DescribeAlarmHistoryResponse (Maybe Text)

The token that marks the start of the next batch of returned results.

describeAlarmHistoryResponse_httpStatus :: Lens' DescribeAlarmHistoryResponse Int

The response's http status code.