Network.AWS.CloudWatch.DescribeAlarmsForMetric

Description

Retrieves the alarms for the specified metric. To filter the results, specify a statistic, period, or unit.

This operation retrieves only standard alarms that are based on the specified metric. It does not return alarms based on math expressions that use the specified metric, or composite alarms that use the specified metric.

Creating a Request

data DescribeAlarmsForMetric

See: newDescribeAlarmsForMetric smart constructor.

Constructors

DescribeAlarmsForMetric'

Fields

extendedStatistic :: Maybe Text

The percentile statistic for the metric. Specify a value between p0.0 and p100.

unit :: Maybe StandardUnit

The unit for the metric.

statistic :: Maybe Statistic

The statistic for the metric, other than percentiles. For percentile statistics, use ExtendedStatistics.

dimensions :: Maybe [Dimension]

The dimensions associated with the metric. If the metric has any associated dimensions, you must specify them in order for the call to succeed.

period :: Maybe Natural

The period, in seconds, over which the statistic is applied.

metricName :: Text

The name of the metric.

namespace :: Text

The namespace of the metric.

newDescribeAlarmsForMetric

Arguments

:: Text
$sel:metricName:DescribeAlarmsForMetric'

-> Text
$sel:namespace:DescribeAlarmsForMetric'

-> DescribeAlarmsForMetric

Create a value of DescribeAlarmsForMetric with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:extendedStatistic:DescribeAlarmsForMetric', describeAlarmsForMetric_extendedStatistic - The percentile statistic for the metric. Specify a value between p0.0 and p100.

$sel:unit:DescribeAlarmsForMetric', describeAlarmsForMetric_unit - The unit for the metric.

$sel:statistic:DescribeAlarmsForMetric', describeAlarmsForMetric_statistic - The statistic for the metric, other than percentiles. For percentile statistics, use ExtendedStatistics.

$sel:dimensions:DescribeAlarmsForMetric', describeAlarmsForMetric_dimensions - The dimensions associated with the metric. If the metric has any associated dimensions, you must specify them in order for the call to succeed.

$sel:period:DescribeAlarmsForMetric', describeAlarmsForMetric_period - The period, in seconds, over which the statistic is applied.

$sel:metricName:DescribeAlarmsForMetric', describeAlarmsForMetric_metricName - The name of the metric.

$sel:namespace:DescribeAlarmsForMetric', describeAlarmsForMetric_namespace - The namespace of the metric.

Request Lenses

describeAlarmsForMetric_extendedStatistic :: Lens' DescribeAlarmsForMetric (Maybe Text)

The percentile statistic for the metric. Specify a value between p0.0 and p100.

describeAlarmsForMetric_unit :: Lens' DescribeAlarmsForMetric (Maybe StandardUnit)

The unit for the metric.

describeAlarmsForMetric_statistic :: Lens' DescribeAlarmsForMetric (Maybe Statistic)

The statistic for the metric, other than percentiles. For percentile statistics, use ExtendedStatistics.

describeAlarmsForMetric_dimensions :: Lens' DescribeAlarmsForMetric (Maybe [Dimension])

The dimensions associated with the metric. If the metric has any associated dimensions, you must specify them in order for the call to succeed.

describeAlarmsForMetric_period :: Lens' DescribeAlarmsForMetric (Maybe Natural)

The period, in seconds, over which the statistic is applied.

describeAlarmsForMetric_metricName :: Lens' DescribeAlarmsForMetric Text

The name of the metric.

describeAlarmsForMetric_namespace :: Lens' DescribeAlarmsForMetric Text

The namespace of the metric.

Destructuring the Response

data DescribeAlarmsForMetricResponse

See: newDescribeAlarmsForMetricResponse smart constructor.

Constructors

DescribeAlarmsForMetricResponse'

Fields

metricAlarms :: Maybe [MetricAlarm]

The information for each alarm with the specified metric.

httpStatus :: Int

The response's http status code.

newDescribeAlarmsForMetricResponse

Arguments

:: Int
$sel:httpStatus:DescribeAlarmsForMetricResponse'

-> DescribeAlarmsForMetricResponse

Create a value of DescribeAlarmsForMetricResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metricAlarms:DescribeAlarmsForMetricResponse', describeAlarmsForMetricResponse_metricAlarms - The information for each alarm with the specified metric.

$sel:httpStatus:DescribeAlarmsForMetricResponse', describeAlarmsForMetricResponse_httpStatus - The response's http status code.

Response Lenses

describeAlarmsForMetricResponse_metricAlarms :: Lens' DescribeAlarmsForMetricResponse (Maybe [MetricAlarm])

The information for each alarm with the specified metric.

describeAlarmsForMetricResponse_httpStatus :: Lens' DescribeAlarmsForMetricResponse Int

The response's http status code.