Network.AWS.CloudWatch.DisableAlarmActions

Description

Disables the actions for the specified alarms. When an alarm's actions are disabled, the alarm actions do not execute when the alarm state changes.

Creating a Request

data DisableAlarmActions

See: newDisableAlarmActions smart constructor.

Constructors

DisableAlarmActions'

Fields

alarmNames :: [Text]

The names of the alarms.

newDisableAlarmActions :: DisableAlarmActions

Create a value of DisableAlarmActions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:alarmNames:DisableAlarmActions', disableAlarmActions_alarmNames - The names of the alarms.

Request Lenses

disableAlarmActions_alarmNames :: Lens' DisableAlarmActions [Text]

The names of the alarms.

Destructuring the Response

data DisableAlarmActionsResponse

See: newDisableAlarmActionsResponse smart constructor.

Constructors

DisableAlarmActionsResponse'

newDisableAlarmActionsResponse :: DisableAlarmActionsResponse

Create a value of DisableAlarmActionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.