Network.AWS.CloudWatch.GetMetricWidgetImage

Description

You can use the GetMetricWidgetImage API to retrieve a snapshot graph of one or more Amazon CloudWatch metrics as a bitmap image. You can then embed this image into your services and products, such as wiki pages, reports, and documents. You could also retrieve images regularly, such as every minute, and create your own custom live dashboard.

The graph you retrieve can include all CloudWatch metric graph features, including metric math and horizontal and vertical annotations.

There is a limit of 20 transactions per second for this API. Each GetMetricWidgetImage action has the following limits:

As many as 100 metrics in the graph.

Up to 100 KB uncompressed payload.

Creating a Request

data GetMetricWidgetImage

See: newGetMetricWidgetImage smart constructor.

Constructors

GetMetricWidgetImage'

Fields

outputFormat :: Maybe Text

The format of the resulting image. Only PNG images are supported.

The default is png. If you specify png, the API returns an HTTP response with the content-type set to text/xml. The image data is in a MetricWidgetImage field. For example:

 <GetMetricWidgetImageResponse xmlns=<URLstring>>

 <GetMetricWidgetImageResult>

 <MetricWidgetImage>

 iVBORw0KGgoAAAANSUhEUgAAAlgAAAGQEAYAAAAip...

 </MetricWidgetImage>

 </GetMetricWidgetImageResult>

 <ResponseMetadata>

 <RequestId>6f0d4192-4d42-11e8-82c1-f539a07e0e3b</RequestId>

 </ResponseMetadata>

</GetMetricWidgetImageResponse>

The image/png setting is intended only for custom HTTP requests. For most use cases, and all actions using an Amazon Web Services SDK, you should use png. If you specify image/png, the HTTP response has a content-type set to image/png, and the body of the response is a PNG image.

metricWidget :: Text

A JSON string that defines the bitmap graph to be retrieved. The string includes the metrics to include in the graph, statistics, annotations, title, axis limits, and so on. You can include only one MetricWidget parameter in each GetMetricWidgetImage call.

For more information about the syntax of MetricWidget see GetMetricWidgetImage: Metric Widget Structure and Syntax.

If any metric on the graph could not load all the requested data points, an orange triangle with an exclamation point appears next to the graph legend.

newGetMetricWidgetImage

Arguments

:: Text
$sel:metricWidget:GetMetricWidgetImage'

-> GetMetricWidgetImage

Create a value of GetMetricWidgetImage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:outputFormat:GetMetricWidgetImage', getMetricWidgetImage_outputFormat - The format of the resulting image. Only PNG images are supported.

The default is png. If you specify png, the API returns an HTTP response with the content-type set to text/xml. The image data is in a MetricWidgetImage field. For example:

 <GetMetricWidgetImageResponse xmlns=<URLstring>>

 <GetMetricWidgetImageResult>

 <MetricWidgetImage>

 iVBORw0KGgoAAAANSUhEUgAAAlgAAAGQEAYAAAAip...

 </MetricWidgetImage>

 </GetMetricWidgetImageResult>

 <ResponseMetadata>

 <RequestId>6f0d4192-4d42-11e8-82c1-f539a07e0e3b</RequestId>

 </ResponseMetadata>

</GetMetricWidgetImageResponse>

The image/png setting is intended only for custom HTTP requests. For most use cases, and all actions using an Amazon Web Services SDK, you should use png. If you specify image/png, the HTTP response has a content-type set to image/png, and the body of the response is a PNG image.

$sel:metricWidget:GetMetricWidgetImage', getMetricWidgetImage_metricWidget - A JSON string that defines the bitmap graph to be retrieved. The string includes the metrics to include in the graph, statistics, annotations, title, axis limits, and so on. You can include only one MetricWidget parameter in each GetMetricWidgetImage call.

For more information about the syntax of MetricWidget see GetMetricWidgetImage: Metric Widget Structure and Syntax.

If any metric on the graph could not load all the requested data points, an orange triangle with an exclamation point appears next to the graph legend.

Request Lenses

getMetricWidgetImage_outputFormat :: Lens' GetMetricWidgetImage (Maybe Text)

The format of the resulting image. Only PNG images are supported.

The default is png. If you specify png, the API returns an HTTP response with the content-type set to text/xml. The image data is in a MetricWidgetImage field. For example:

 <GetMetricWidgetImageResponse xmlns=<URLstring>>

 <GetMetricWidgetImageResult>

 <MetricWidgetImage>

 iVBORw0KGgoAAAANSUhEUgAAAlgAAAGQEAYAAAAip...

 </MetricWidgetImage>

 </GetMetricWidgetImageResult>

 <ResponseMetadata>

 <RequestId>6f0d4192-4d42-11e8-82c1-f539a07e0e3b</RequestId>

 </ResponseMetadata>

</GetMetricWidgetImageResponse>

The image/png setting is intended only for custom HTTP requests. For most use cases, and all actions using an Amazon Web Services SDK, you should use png. If you specify image/png, the HTTP response has a content-type set to image/png, and the body of the response is a PNG image.

getMetricWidgetImage_metricWidget :: Lens' GetMetricWidgetImage Text

A JSON string that defines the bitmap graph to be retrieved. The string includes the metrics to include in the graph, statistics, annotations, title, axis limits, and so on. You can include only one MetricWidget parameter in each GetMetricWidgetImage call.

For more information about the syntax of MetricWidget see GetMetricWidgetImage: Metric Widget Structure and Syntax.

If any metric on the graph could not load all the requested data points, an orange triangle with an exclamation point appears next to the graph legend.

Destructuring the Response

data GetMetricWidgetImageResponse

See: newGetMetricWidgetImageResponse smart constructor.

Constructors

GetMetricWidgetImageResponse'

Fields

metricWidgetImage :: Maybe Base64

The image of the graph, in the output format specified. The output is base64-encoded.

httpStatus :: Int

The response's http status code.

newGetMetricWidgetImageResponse

Arguments

:: Int
$sel:httpStatus:GetMetricWidgetImageResponse'

-> GetMetricWidgetImageResponse

Create a value of GetMetricWidgetImageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metricWidgetImage:GetMetricWidgetImageResponse', getMetricWidgetImageResponse_metricWidgetImage - The image of the graph, in the output format specified. The output is base64-encoded.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:httpStatus:GetMetricWidgetImageResponse', getMetricWidgetImageResponse_httpStatus - The response's http status code.

Response Lenses

getMetricWidgetImageResponse_metricWidgetImage :: Lens' GetMetricWidgetImageResponse (Maybe ByteString)

The image of the graph, in the output format specified. The output is base64-encoded.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

getMetricWidgetImageResponse_httpStatus :: Lens' GetMetricWidgetImageResponse Int

The response's http status code.