Network.AWS.CloudWatch.ListDashboards

Description

Returns a list of the dashboards for your account. If you include DashboardNamePrefix, only those dashboards with names starting with the prefix are listed. Otherwise, all dashboards in your account are listed.

ListDashboards returns up to 1000 results on one page. If there are more than 1000 dashboards, you can call ListDashboards again and include the value you received for NextToken in the first call, to receive the next 1000 results.

This operation returns paginated results.

Creating a Request

data ListDashboards

See: newListDashboards smart constructor.

Constructors

ListDashboards'

Fields

dashboardNamePrefix :: Maybe Text

If you specify this parameter, only the dashboards with names starting with the specified string are listed. The maximum length is 255, and valid characters are A-Z, a-z, 0-9, ".", "-", and "_".

nextToken :: Maybe Text

The token returned by a previous call to indicate that there is more data available.

newListDashboards :: ListDashboards

Create a value of ListDashboards with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dashboardNamePrefix:ListDashboards', listDashboards_dashboardNamePrefix - If you specify this parameter, only the dashboards with names starting with the specified string are listed. The maximum length is 255, and valid characters are A-Z, a-z, 0-9, ".", "-", and "_".

$sel:nextToken:ListDashboards', listDashboards_nextToken - The token returned by a previous call to indicate that there is more data available.

Request Lenses

listDashboards_dashboardNamePrefix :: Lens' ListDashboards (Maybe Text)

If you specify this parameter, only the dashboards with names starting with the specified string are listed. The maximum length is 255, and valid characters are A-Z, a-z, 0-9, ".", "-", and "_".

listDashboards_nextToken :: Lens' ListDashboards (Maybe Text)

The token returned by a previous call to indicate that there is more data available.

Destructuring the Response

data ListDashboardsResponse

See: newListDashboardsResponse smart constructor.

Constructors

ListDashboardsResponse'

Fields

dashboardEntries :: Maybe [DashboardEntry]

The list of matching dashboards.

nextToken :: Maybe Text

The token that marks the start of the next batch of returned results.

httpStatus :: Int

The response's http status code.

newListDashboardsResponse

Arguments

:: Int
$sel:httpStatus:ListDashboardsResponse'

-> ListDashboardsResponse

Create a value of ListDashboardsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dashboardEntries:ListDashboardsResponse', listDashboardsResponse_dashboardEntries - The list of matching dashboards.

$sel:nextToken:ListDashboards', listDashboardsResponse_nextToken - The token that marks the start of the next batch of returned results.

$sel:httpStatus:ListDashboardsResponse', listDashboardsResponse_httpStatus - The response's http status code.

Response Lenses

listDashboardsResponse_dashboardEntries :: Lens' ListDashboardsResponse (Maybe [DashboardEntry])

The list of matching dashboards.

listDashboardsResponse_nextToken :: Lens' ListDashboardsResponse (Maybe Text)

The token that marks the start of the next batch of returned results.

listDashboardsResponse_httpStatus :: Lens' ListDashboardsResponse Int

The response's http status code.