Network.AWS.CloudWatch.ListMetricStreams

Description

Returns a list of metric streams in this account.

Creating a Request

data ListMetricStreams

See: newListMetricStreams smart constructor.

Constructors

ListMetricStreams'

Fields

nextToken :: Maybe Text

Include this value, if it was returned by the previous call, to get the next set of metric streams.

maxResults :: Maybe Natural

The maximum number of results to return in one operation.

newListMetricStreams :: ListMetricStreams

Create a value of ListMetricStreams with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListMetricStreams', listMetricStreams_nextToken - Include this value, if it was returned by the previous call, to get the next set of metric streams.

$sel:maxResults:ListMetricStreams', listMetricStreams_maxResults - The maximum number of results to return in one operation.

Request Lenses

listMetricStreams_nextToken :: Lens' ListMetricStreams (Maybe Text)

Include this value, if it was returned by the previous call, to get the next set of metric streams.

listMetricStreams_maxResults :: Lens' ListMetricStreams (Maybe Natural)

The maximum number of results to return in one operation.

Destructuring the Response

data ListMetricStreamsResponse

See: newListMetricStreamsResponse smart constructor.

Constructors

ListMetricStreamsResponse'

Fields

nextToken :: Maybe Text

The token that marks the start of the next batch of returned results. You can use this token in a subsequent operation to get the next batch of results.

entries :: Maybe [MetricStreamEntry]

The array of metric stream information.

httpStatus :: Int

The response's http status code.

newListMetricStreamsResponse

Arguments

:: Int
$sel:httpStatus:ListMetricStreamsResponse'

-> ListMetricStreamsResponse

Create a value of ListMetricStreamsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListMetricStreams', listMetricStreamsResponse_nextToken - The token that marks the start of the next batch of returned results. You can use this token in a subsequent operation to get the next batch of results.

$sel:entries:ListMetricStreamsResponse', listMetricStreamsResponse_entries - The array of metric stream information.

$sel:httpStatus:ListMetricStreamsResponse', listMetricStreamsResponse_httpStatus - The response's http status code.

Response Lenses

listMetricStreamsResponse_nextToken :: Lens' ListMetricStreamsResponse (Maybe Text)

The token that marks the start of the next batch of returned results. You can use this token in a subsequent operation to get the next batch of results.

listMetricStreamsResponse_entries :: Lens' ListMetricStreamsResponse (Maybe [MetricStreamEntry])

The array of metric stream information.

listMetricStreamsResponse_httpStatus :: Lens' ListMetricStreamsResponse Int

The response's http status code.