Network.AWS.CloudWatch.PutMetricStream

Description

Creates or updates a metric stream. Metric streams can automatically stream CloudWatch metrics to Amazon Web Services destinations including Amazon S3 and to many third-party solutions.

For more information, see Using Metric Streams.

To create a metric stream, you must be logged on to an account that has the iam:PassRole permission and either the CloudWatchFullAccess policy or the cloudwatch:PutMetricStream permission.

When you create or update a metric stream, you choose one of the following:

Stream metrics from all metric namespaces in the account.

Stream metrics from all metric namespaces in the account, except for the namespaces that you list in ExcludeFilters.

Stream metrics from only the metric namespaces that you list in IncludeFilters.

When you use PutMetricStream to create a new metric stream, the stream is created in the running state. If you use it to update an existing stream, the state of the stream is not changed.

Creating a Request

data PutMetricStream

See: newPutMetricStream smart constructor.

Constructors

PutMetricStream'

Fields

includeFilters :: Maybe [MetricStreamFilter]

If you specify this parameter, the stream sends only the metrics from the metric namespaces that you specify here.

You cannot include IncludeFilters and ExcludeFilters in the same operation.

excludeFilters :: Maybe [MetricStreamFilter]

If you specify this parameter, the stream sends metrics from all metric namespaces except for the namespaces that you specify here.

You cannot include ExcludeFilters and IncludeFilters in the same operation.

tags :: Maybe [Tag]

A list of key-value pairs to associate with the metric stream. You can associate as many as 50 tags with a metric stream.

Tags can help you organize and categorize your resources. You can also use them to scope user permissions by granting a user permission to access or change only resources with certain tag values.

You can use this parameter only when you are creating a new metric stream. If you are using this operation to update an existing metric stream, any tags you specify in this parameter are ignored. To change the tags of an existing metric stream, use TagResource or UntagResource.

name :: Text

If you are creating a new metric stream, this is the name for the new stream. The name must be different than the names of other metric streams in this account and Region.

If you are updating a metric stream, specify the name of that stream here.

Valid characters are A-Z, a-z, 0-9, "-" and "_".

firehoseArn :: Text

The ARN of the Amazon Kinesis Firehose delivery stream to use for this metric stream. This Amazon Kinesis Firehose delivery stream must already exist and must be in the same account as the metric stream.

roleArn :: Text

The ARN of an IAM role that this metric stream will use to access Amazon Kinesis Firehose resources. This IAM role must already exist and must be in the same account as the metric stream. This IAM role must include the following permissions:

firehose:PutRecord

firehose:PutRecordBatch

outputFormat :: MetricStreamOutputFormat

The output format for the stream. Valid values are json and opentelemetry0.7. For more information about metric stream output formats, see Metric streams output formats.

newPutMetricStream

Arguments

:: Text
$sel:name:PutMetricStream'

-> Text
$sel:firehoseArn:PutMetricStream'

-> Text
$sel:roleArn:PutMetricStream'

-> MetricStreamOutputFormat
$sel:outputFormat:PutMetricStream'

-> PutMetricStream

Create a value of PutMetricStream with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:includeFilters:PutMetricStream', putMetricStream_includeFilters - If you specify this parameter, the stream sends only the metrics from the metric namespaces that you specify here.

You cannot include IncludeFilters and ExcludeFilters in the same operation.

$sel:excludeFilters:PutMetricStream', putMetricStream_excludeFilters - If you specify this parameter, the stream sends metrics from all metric namespaces except for the namespaces that you specify here.

You cannot include ExcludeFilters and IncludeFilters in the same operation.

$sel:tags:PutMetricStream', putMetricStream_tags - A list of key-value pairs to associate with the metric stream. You can associate as many as 50 tags with a metric stream.

Tags can help you organize and categorize your resources. You can also use them to scope user permissions by granting a user permission to access or change only resources with certain tag values.

You can use this parameter only when you are creating a new metric stream. If you are using this operation to update an existing metric stream, any tags you specify in this parameter are ignored. To change the tags of an existing metric stream, use TagResource or UntagResource.

$sel:name:PutMetricStream', putMetricStream_name - If you are creating a new metric stream, this is the name for the new stream. The name must be different than the names of other metric streams in this account and Region.

If you are updating a metric stream, specify the name of that stream here.

Valid characters are A-Z, a-z, 0-9, "-" and "_".

$sel:firehoseArn:PutMetricStream', putMetricStream_firehoseArn - The ARN of the Amazon Kinesis Firehose delivery stream to use for this metric stream. This Amazon Kinesis Firehose delivery stream must already exist and must be in the same account as the metric stream.

$sel:roleArn:PutMetricStream', putMetricStream_roleArn - The ARN of an IAM role that this metric stream will use to access Amazon Kinesis Firehose resources. This IAM role must already exist and must be in the same account as the metric stream. This IAM role must include the following permissions:

firehose:PutRecord

firehose:PutRecordBatch

$sel:outputFormat:PutMetricStream', putMetricStream_outputFormat - The output format for the stream. Valid values are json and opentelemetry0.7. For more information about metric stream output formats, see Metric streams output formats.

Request Lenses

putMetricStream_includeFilters :: Lens' PutMetricStream (Maybe [MetricStreamFilter])

If you specify this parameter, the stream sends only the metrics from the metric namespaces that you specify here.

You cannot include IncludeFilters and ExcludeFilters in the same operation.

putMetricStream_excludeFilters :: Lens' PutMetricStream (Maybe [MetricStreamFilter])

If you specify this parameter, the stream sends metrics from all metric namespaces except for the namespaces that you specify here.

You cannot include ExcludeFilters and IncludeFilters in the same operation.

putMetricStream_tags :: Lens' PutMetricStream (Maybe [Tag])

A list of key-value pairs to associate with the metric stream. You can associate as many as 50 tags with a metric stream.

Tags can help you organize and categorize your resources. You can also use them to scope user permissions by granting a user permission to access or change only resources with certain tag values.

You can use this parameter only when you are creating a new metric stream. If you are using this operation to update an existing metric stream, any tags you specify in this parameter are ignored. To change the tags of an existing metric stream, use TagResource or UntagResource.

putMetricStream_name :: Lens' PutMetricStream Text

If you are creating a new metric stream, this is the name for the new stream. The name must be different than the names of other metric streams in this account and Region.

If you are updating a metric stream, specify the name of that stream here.

Valid characters are A-Z, a-z, 0-9, "-" and "_".

putMetricStream_firehoseArn :: Lens' PutMetricStream Text

The ARN of the Amazon Kinesis Firehose delivery stream to use for this metric stream. This Amazon Kinesis Firehose delivery stream must already exist and must be in the same account as the metric stream.

putMetricStream_roleArn :: Lens' PutMetricStream Text

The ARN of an IAM role that this metric stream will use to access Amazon Kinesis Firehose resources. This IAM role must already exist and must be in the same account as the metric stream. This IAM role must include the following permissions:

firehose:PutRecord

firehose:PutRecordBatch

putMetricStream_outputFormat :: Lens' PutMetricStream MetricStreamOutputFormat

The output format for the stream. Valid values are json and opentelemetry0.7. For more information about metric stream output formats, see Metric streams output formats.

Destructuring the Response

data PutMetricStreamResponse

See: newPutMetricStreamResponse smart constructor.

Constructors

PutMetricStreamResponse'

Fields

arn :: Maybe Text

The ARN of the metric stream.

httpStatus :: Int

The response's http status code.

newPutMetricStreamResponse

Arguments

:: Int
$sel:httpStatus:PutMetricStreamResponse'

-> PutMetricStreamResponse

Create a value of PutMetricStreamResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:PutMetricStreamResponse', putMetricStreamResponse_arn - The ARN of the metric stream.

$sel:httpStatus:PutMetricStreamResponse', putMetricStreamResponse_httpStatus - The response's http status code.

Response Lenses

putMetricStreamResponse_arn :: Lens' PutMetricStreamResponse (Maybe Text)

The ARN of the metric stream.

putMetricStreamResponse_httpStatus :: Lens' PutMetricStreamResponse Int

The response's http status code.