Network.AWS.CloudWatch.SetAlarmState

Description

Temporarily sets the state of an alarm for testing purposes. When the updated state differs from the previous value, the action configured for the appropriate state is invoked. For example, if your alarm is configured to send an Amazon SNS message when an alarm is triggered, temporarily changing the alarm state to ALARM sends an SNS message.

Metric alarms returns to their actual state quickly, often within seconds. Because the metric alarm state change happens quickly, it is typically only visible in the alarm's History tab in the Amazon CloudWatch console or through DescribeAlarmHistory.

If you use SetAlarmState on a composite alarm, the composite alarm is not guaranteed to return to its actual state. It returns to its actual state only once any of its children alarms change state. It is also reevaluated if you update its configuration.

If an alarm triggers EC2 Auto Scaling policies or application Auto Scaling policies, you must include information in the StateReasonData parameter to enable the policy to take the correct action.

Creating a Request

data SetAlarmState

See: newSetAlarmState smart constructor.

Constructors

SetAlarmState'

Fields

stateReasonData :: Maybe Text

The reason that this alarm is set to this specific state, in JSON format.

For SNS or EC2 alarm actions, this is just informational. But for EC2 Auto Scaling or application Auto Scaling alarm actions, the Auto Scaling policy uses the information in this field to take the correct action.

alarmName :: Text

The name of the alarm.

stateValue :: StateValue

The value of the state.

stateReason :: Text

The reason that this alarm is set to this specific state, in text format.

newSetAlarmState

Arguments

:: Text
$sel:alarmName:SetAlarmState'

-> StateValue
$sel:stateValue:SetAlarmState'

-> Text
$sel:stateReason:SetAlarmState'

-> SetAlarmState

Create a value of SetAlarmState with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stateReasonData:SetAlarmState', setAlarmState_stateReasonData - The reason that this alarm is set to this specific state, in JSON format.

For SNS or EC2 alarm actions, this is just informational. But for EC2 Auto Scaling or application Auto Scaling alarm actions, the Auto Scaling policy uses the information in this field to take the correct action.

$sel:alarmName:SetAlarmState', setAlarmState_alarmName - The name of the alarm.

$sel:stateValue:SetAlarmState', setAlarmState_stateValue - The value of the state.

$sel:stateReason:SetAlarmState', setAlarmState_stateReason - The reason that this alarm is set to this specific state, in text format.

Request Lenses

setAlarmState_stateReasonData :: Lens' SetAlarmState (Maybe Text)

The reason that this alarm is set to this specific state, in JSON format.

For SNS or EC2 alarm actions, this is just informational. But for EC2 Auto Scaling or application Auto Scaling alarm actions, the Auto Scaling policy uses the information in this field to take the correct action.

setAlarmState_alarmName :: Lens' SetAlarmState Text

The name of the alarm.

setAlarmState_stateValue :: Lens' SetAlarmState StateValue

The value of the state.

setAlarmState_stateReason :: Lens' SetAlarmState Text

The reason that this alarm is set to this specific state, in text format.

Destructuring the Response

data SetAlarmStateResponse

See: newSetAlarmStateResponse smart constructor.

Constructors

SetAlarmStateResponse'

newSetAlarmStateResponse :: SetAlarmStateResponse

Create a value of SetAlarmStateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.