Network.AWS.CloudWatch.Types.CompositeAlarm

Description

Documentation

data CompositeAlarm

The details about a composite alarm.

See: newCompositeAlarm smart constructor.

Constructors

CompositeAlarm'

Fields

alarmName :: Maybe Text

The name of the alarm.

stateUpdatedTimestamp :: Maybe ISO8601

The time stamp of the last update to the alarm state.

alarmDescription :: Maybe Text

The description of the alarm.

alarmRule :: Maybe Text

The rule that this alarm uses to evaluate its alarm state.

oKActions :: Maybe [Text]

The actions to execute when this alarm transitions to the OK state from any other state. Each action is specified as an Amazon Resource Name (ARN).

stateValue :: Maybe StateValue

The state value for the alarm.

alarmConfigurationUpdatedTimestamp :: Maybe ISO8601

The time stamp of the last update to the alarm configuration.

actionsEnabled :: Maybe Bool

Indicates whether actions should be executed during any changes to the alarm state.

insufficientDataActions :: Maybe [Text]

The actions to execute when this alarm transitions to the INSUFFICIENT_DATA state from any other state. Each action is specified as an Amazon Resource Name (ARN).

stateReason :: Maybe Text

An explanation for the alarm state, in text format.

stateReasonData :: Maybe Text

An explanation for the alarm state, in JSON format.

alarmArn :: Maybe Text

The Amazon Resource Name (ARN) of the alarm.

alarmActions :: Maybe [Text]

The actions to execute when this alarm transitions to the ALARM state from any other state. Each action is specified as an Amazon Resource Name (ARN).

newCompositeAlarm :: CompositeAlarm

Create a value of CompositeAlarm with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:alarmName:CompositeAlarm', compositeAlarm_alarmName - The name of the alarm.

$sel:stateUpdatedTimestamp:CompositeAlarm', compositeAlarm_stateUpdatedTimestamp - The time stamp of the last update to the alarm state.

$sel:alarmDescription:CompositeAlarm', compositeAlarm_alarmDescription - The description of the alarm.

$sel:alarmRule:CompositeAlarm', compositeAlarm_alarmRule - The rule that this alarm uses to evaluate its alarm state.

$sel:oKActions:CompositeAlarm', compositeAlarm_oKActions - The actions to execute when this alarm transitions to the OK state from any other state. Each action is specified as an Amazon Resource Name (ARN).

$sel:stateValue:CompositeAlarm', compositeAlarm_stateValue - The state value for the alarm.

$sel:alarmConfigurationUpdatedTimestamp:CompositeAlarm', compositeAlarm_alarmConfigurationUpdatedTimestamp - The time stamp of the last update to the alarm configuration.

$sel:actionsEnabled:CompositeAlarm', compositeAlarm_actionsEnabled - Indicates whether actions should be executed during any changes to the alarm state.

$sel:insufficientDataActions:CompositeAlarm', compositeAlarm_insufficientDataActions - The actions to execute when this alarm transitions to the INSUFFICIENT_DATA state from any other state. Each action is specified as an Amazon Resource Name (ARN).

$sel:stateReason:CompositeAlarm', compositeAlarm_stateReason - An explanation for the alarm state, in text format.

$sel:stateReasonData:CompositeAlarm', compositeAlarm_stateReasonData - An explanation for the alarm state, in JSON format.

$sel:alarmArn:CompositeAlarm', compositeAlarm_alarmArn - The Amazon Resource Name (ARN) of the alarm.

$sel:alarmActions:CompositeAlarm', compositeAlarm_alarmActions - The actions to execute when this alarm transitions to the ALARM state from any other state. Each action is specified as an Amazon Resource Name (ARN).

compositeAlarm_alarmName :: Lens' CompositeAlarm (Maybe Text)

The name of the alarm.

compositeAlarm_stateUpdatedTimestamp :: Lens' CompositeAlarm (Maybe UTCTime)

The time stamp of the last update to the alarm state.

compositeAlarm_alarmDescription :: Lens' CompositeAlarm (Maybe Text)

The description of the alarm.

compositeAlarm_alarmRule :: Lens' CompositeAlarm (Maybe Text)

The rule that this alarm uses to evaluate its alarm state.

compositeAlarm_oKActions :: Lens' CompositeAlarm (Maybe [Text])

The actions to execute when this alarm transitions to the OK state from any other state. Each action is specified as an Amazon Resource Name (ARN).

compositeAlarm_stateValue :: Lens' CompositeAlarm (Maybe StateValue)

The state value for the alarm.

compositeAlarm_alarmConfigurationUpdatedTimestamp :: Lens' CompositeAlarm (Maybe UTCTime)

The time stamp of the last update to the alarm configuration.

compositeAlarm_actionsEnabled :: Lens' CompositeAlarm (Maybe Bool)

Indicates whether actions should be executed during any changes to the alarm state.

compositeAlarm_insufficientDataActions :: Lens' CompositeAlarm (Maybe [Text])

The actions to execute when this alarm transitions to the INSUFFICIENT_DATA state from any other state. Each action is specified as an Amazon Resource Name (ARN).

compositeAlarm_stateReason :: Lens' CompositeAlarm (Maybe Text)

An explanation for the alarm state, in text format.

compositeAlarm_stateReasonData :: Lens' CompositeAlarm (Maybe Text)

An explanation for the alarm state, in JSON format.

compositeAlarm_alarmArn :: Lens' CompositeAlarm (Maybe Text)

The Amazon Resource Name (ARN) of the alarm.

compositeAlarm_alarmActions :: Lens' CompositeAlarm (Maybe [Text])

The actions to execute when this alarm transitions to the ALARM state from any other state. Each action is specified as an Amazon Resource Name (ARN).