Network.AWS.CloudWatch.Types.InsightRuleMetricDatapoint

Description

Documentation

data InsightRuleMetricDatapoint

One data point from the metric time series returned in a Contributor Insights rule report.

For more information, see GetInsightRuleReport.

See: newInsightRuleMetricDatapoint smart constructor.

Constructors

InsightRuleMetricDatapoint'

Fields

minimum :: Maybe Double

The minimum value from a single contributor during the time period represented by that data point.

This statistic is returned only if you included it in the Metrics array in your request.

sum :: Maybe Double

The sum of the values from all contributors during the time period represented by that data point.

This statistic is returned only if you included it in the Metrics array in your request.

sampleCount :: Maybe Double

The number of occurrences that matched the rule during this data point.

This statistic is returned only if you included it in the Metrics array in your request.

maxContributorValue :: Maybe Double

The maximum value provided by one contributor during this timestamp. Each timestamp is evaluated separately, so the identity of the max contributor could be different for each timestamp.

This statistic is returned only if you included it in the Metrics array in your request.

average :: Maybe Double

The average value from all contributors during the time period represented by that data point.

This statistic is returned only if you included it in the Metrics array in your request.

uniqueContributors :: Maybe Double

The number of unique contributors who published data during this timestamp.

This statistic is returned only if you included it in the Metrics array in your request.

maximum :: Maybe Double

The maximum value from a single occurence from a single contributor during the time period represented by that data point.

This statistic is returned only if you included it in the Metrics array in your request.

timestamp :: ISO8601

The timestamp of the data point.

newInsightRuleMetricDatapoint

Arguments

:: UTCTime
$sel:timestamp:InsightRuleMetricDatapoint'

-> InsightRuleMetricDatapoint

Create a value of InsightRuleMetricDatapoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:minimum:InsightRuleMetricDatapoint', insightRuleMetricDatapoint_minimum - The minimum value from a single contributor during the time period represented by that data point.

This statistic is returned only if you included it in the Metrics array in your request.

$sel:sum:InsightRuleMetricDatapoint', insightRuleMetricDatapoint_sum - The sum of the values from all contributors during the time period represented by that data point.

This statistic is returned only if you included it in the Metrics array in your request.

$sel:sampleCount:InsightRuleMetricDatapoint', insightRuleMetricDatapoint_sampleCount - The number of occurrences that matched the rule during this data point.

This statistic is returned only if you included it in the Metrics array in your request.

$sel:maxContributorValue:InsightRuleMetricDatapoint', insightRuleMetricDatapoint_maxContributorValue - The maximum value provided by one contributor during this timestamp. Each timestamp is evaluated separately, so the identity of the max contributor could be different for each timestamp.

This statistic is returned only if you included it in the Metrics array in your request.

$sel:average:InsightRuleMetricDatapoint', insightRuleMetricDatapoint_average - The average value from all contributors during the time period represented by that data point.

This statistic is returned only if you included it in the Metrics array in your request.

$sel:uniqueContributors:InsightRuleMetricDatapoint', insightRuleMetricDatapoint_uniqueContributors - The number of unique contributors who published data during this timestamp.

This statistic is returned only if you included it in the Metrics array in your request.

$sel:maximum:InsightRuleMetricDatapoint', insightRuleMetricDatapoint_maximum - The maximum value from a single occurence from a single contributor during the time period represented by that data point.

This statistic is returned only if you included it in the Metrics array in your request.

$sel:timestamp:InsightRuleMetricDatapoint', insightRuleMetricDatapoint_timestamp - The timestamp of the data point.

insightRuleMetricDatapoint_minimum :: Lens' InsightRuleMetricDatapoint (Maybe Double)

The minimum value from a single contributor during the time period represented by that data point.

This statistic is returned only if you included it in the Metrics array in your request.

insightRuleMetricDatapoint_sum :: Lens' InsightRuleMetricDatapoint (Maybe Double)

The sum of the values from all contributors during the time period represented by that data point.

This statistic is returned only if you included it in the Metrics array in your request.

insightRuleMetricDatapoint_sampleCount :: Lens' InsightRuleMetricDatapoint (Maybe Double)

The number of occurrences that matched the rule during this data point.

This statistic is returned only if you included it in the Metrics array in your request.

insightRuleMetricDatapoint_maxContributorValue :: Lens' InsightRuleMetricDatapoint (Maybe Double)

The maximum value provided by one contributor during this timestamp. Each timestamp is evaluated separately, so the identity of the max contributor could be different for each timestamp.

This statistic is returned only if you included it in the Metrics array in your request.

insightRuleMetricDatapoint_average :: Lens' InsightRuleMetricDatapoint (Maybe Double)

The average value from all contributors during the time period represented by that data point.

This statistic is returned only if you included it in the Metrics array in your request.

insightRuleMetricDatapoint_uniqueContributors :: Lens' InsightRuleMetricDatapoint (Maybe Double)

The number of unique contributors who published data during this timestamp.

This statistic is returned only if you included it in the Metrics array in your request.

insightRuleMetricDatapoint_maximum :: Lens' InsightRuleMetricDatapoint (Maybe Double)

The maximum value from a single occurence from a single contributor during the time period represented by that data point.

This statistic is returned only if you included it in the Metrics array in your request.

insightRuleMetricDatapoint_timestamp :: Lens' InsightRuleMetricDatapoint UTCTime

The timestamp of the data point.