Network.AWS.CloudWatch.Types.Metric

Description

Documentation

data Metric

Represents a specific metric.

See: newMetric smart constructor.

Constructors

Metric'

Fields

metricName :: Maybe Text

The name of the metric. This is a required field.

namespace :: Maybe Text

The namespace of the metric.

dimensions :: Maybe [Dimension]

The dimensions for the metric.

newMetric :: Metric

Create a value of Metric with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metricName:Metric', metric_metricName - The name of the metric. This is a required field.

$sel:namespace:Metric', metric_namespace - The namespace of the metric.

$sel:dimensions:Metric', metric_dimensions - The dimensions for the metric.

metric_metricName :: Lens' Metric (Maybe Text)

The name of the metric. This is a required field.

metric_namespace :: Lens' Metric (Maybe Text)

The namespace of the metric.

metric_dimensions :: Lens' Metric (Maybe [Dimension])

The dimensions for the metric.