Network.AWS.CloudWatch.Types.MetricAlarm

Description

Documentation

data MetricAlarm

The details about a metric alarm.

See: newMetricAlarm smart constructor.

Constructors

MetricAlarm'

Fields

threshold :: Maybe Double

The value to compare with the specified statistic.

datapointsToAlarm :: Maybe Natural

The number of data points that must be breaching to trigger the alarm.

evaluateLowSampleCountPercentile :: Maybe Text

Used only for alarms based on percentiles. If ignore, the alarm state does not change during periods with too few data points to be statistically significant. If evaluate or this parameter is not used, the alarm is always evaluated and possibly changes state no matter how many data points are available.

comparisonOperator :: Maybe ComparisonOperator

The arithmetic operation to use when comparing the specified statistic and threshold. The specified statistic value is used as the first operand.

extendedStatistic :: Maybe Text

The percentile statistic for the metric associated with the alarm. Specify a value between p0.0 and p100.

thresholdMetricId :: Maybe Text

In an alarm based on an anomaly detection model, this is the ID of the ANOMALY_DETECTION_BAND function used as the threshold for the alarm.

alarmActions :: Maybe [Text]

The actions to execute when this alarm transitions to the ALARM state from any other state. Each action is specified as an Amazon Resource Name (ARN).

unit :: Maybe StandardUnit

The unit of the metric associated with the alarm.

alarmArn :: Maybe Text

The Amazon Resource Name (ARN) of the alarm.

metricName :: Maybe Text

The name of the metric associated with the alarm, if this is an alarm based on a single metric.

stateReason :: Maybe Text

An explanation for the alarm state, in text format.

stateReasonData :: Maybe Text

An explanation for the alarm state, in JSON format.

insufficientDataActions :: Maybe [Text]

The actions to execute when this alarm transitions to the INSUFFICIENT_DATA state from any other state. Each action is specified as an Amazon Resource Name (ARN).

treatMissingData :: Maybe Text

Sets how this alarm is to handle missing data points. If this parameter is omitted, the default behavior of missing is used.

metrics :: Maybe [MetricDataQuery]

An array of MetricDataQuery structures, used in an alarm based on a metric math expression. Each structure either retrieves a metric or performs a math expression. One item in the Metrics array is the math expression that the alarm watches. This expression by designated by having ReturnData set to true.

alarmName :: Maybe Text

The name of the alarm.

stateValue :: Maybe StateValue

The state value for the alarm.

stateUpdatedTimestamp :: Maybe ISO8601

The time stamp of the last update to the alarm state.

oKActions :: Maybe [Text]

The actions to execute when this alarm transitions to the OK state from any other state. Each action is specified as an Amazon Resource Name (ARN).

statistic :: Maybe Statistic

The statistic for the metric associated with the alarm, other than percentile. For percentile statistics, use ExtendedStatistic.

namespace :: Maybe Text

The namespace of the metric associated with the alarm.

dimensions :: Maybe [Dimension]

The dimensions for the metric associated with the alarm.

evaluationPeriods :: Maybe Natural

The number of periods over which data is compared to the specified threshold.

actionsEnabled :: Maybe Bool

Indicates whether actions should be executed during any changes to the alarm state.

alarmConfigurationUpdatedTimestamp :: Maybe ISO8601

The time stamp of the last update to the alarm configuration.

period :: Maybe Natural

The period, in seconds, over which the statistic is applied.

alarmDescription :: Maybe Text

The description of the alarm.

newMetricAlarm :: MetricAlarm

Create a value of MetricAlarm with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:threshold:MetricAlarm', metricAlarm_threshold - The value to compare with the specified statistic.

$sel:datapointsToAlarm:MetricAlarm', metricAlarm_datapointsToAlarm - The number of data points that must be breaching to trigger the alarm.

$sel:evaluateLowSampleCountPercentile:MetricAlarm', metricAlarm_evaluateLowSampleCountPercentile - Used only for alarms based on percentiles. If ignore, the alarm state does not change during periods with too few data points to be statistically significant. If evaluate or this parameter is not used, the alarm is always evaluated and possibly changes state no matter how many data points are available.

$sel:comparisonOperator:MetricAlarm', metricAlarm_comparisonOperator - The arithmetic operation to use when comparing the specified statistic and threshold. The specified statistic value is used as the first operand.

$sel:extendedStatistic:MetricAlarm', metricAlarm_extendedStatistic - The percentile statistic for the metric associated with the alarm. Specify a value between p0.0 and p100.

$sel:thresholdMetricId:MetricAlarm', metricAlarm_thresholdMetricId - In an alarm based on an anomaly detection model, this is the ID of the ANOMALY_DETECTION_BAND function used as the threshold for the alarm.

$sel:alarmActions:MetricAlarm', metricAlarm_alarmActions - The actions to execute when this alarm transitions to the ALARM state from any other state. Each action is specified as an Amazon Resource Name (ARN).

$sel:unit:MetricAlarm', metricAlarm_unit - The unit of the metric associated with the alarm.

$sel:alarmArn:MetricAlarm', metricAlarm_alarmArn - The Amazon Resource Name (ARN) of the alarm.

$sel:metricName:MetricAlarm', metricAlarm_metricName - The name of the metric associated with the alarm, if this is an alarm based on a single metric.

$sel:stateReason:MetricAlarm', metricAlarm_stateReason - An explanation for the alarm state, in text format.

$sel:stateReasonData:MetricAlarm', metricAlarm_stateReasonData - An explanation for the alarm state, in JSON format.

$sel:insufficientDataActions:MetricAlarm', metricAlarm_insufficientDataActions - The actions to execute when this alarm transitions to the INSUFFICIENT_DATA state from any other state. Each action is specified as an Amazon Resource Name (ARN).

$sel:treatMissingData:MetricAlarm', metricAlarm_treatMissingData - Sets how this alarm is to handle missing data points. If this parameter is omitted, the default behavior of missing is used.

$sel:metrics:MetricAlarm', metricAlarm_metrics - An array of MetricDataQuery structures, used in an alarm based on a metric math expression. Each structure either retrieves a metric or performs a math expression. One item in the Metrics array is the math expression that the alarm watches. This expression by designated by having ReturnData set to true.

$sel:alarmName:MetricAlarm', metricAlarm_alarmName - The name of the alarm.

$sel:stateValue:MetricAlarm', metricAlarm_stateValue - The state value for the alarm.

$sel:stateUpdatedTimestamp:MetricAlarm', metricAlarm_stateUpdatedTimestamp - The time stamp of the last update to the alarm state.

$sel:oKActions:MetricAlarm', metricAlarm_oKActions - The actions to execute when this alarm transitions to the OK state from any other state. Each action is specified as an Amazon Resource Name (ARN).

$sel:statistic:MetricAlarm', metricAlarm_statistic - The statistic for the metric associated with the alarm, other than percentile. For percentile statistics, use ExtendedStatistic.

$sel:namespace:MetricAlarm', metricAlarm_namespace - The namespace of the metric associated with the alarm.

$sel:dimensions:MetricAlarm', metricAlarm_dimensions - The dimensions for the metric associated with the alarm.

$sel:evaluationPeriods:MetricAlarm', metricAlarm_evaluationPeriods - The number of periods over which data is compared to the specified threshold.

$sel:actionsEnabled:MetricAlarm', metricAlarm_actionsEnabled - Indicates whether actions should be executed during any changes to the alarm state.

$sel:alarmConfigurationUpdatedTimestamp:MetricAlarm', metricAlarm_alarmConfigurationUpdatedTimestamp - The time stamp of the last update to the alarm configuration.

$sel:period:MetricAlarm', metricAlarm_period - The period, in seconds, over which the statistic is applied.

$sel:alarmDescription:MetricAlarm', metricAlarm_alarmDescription - The description of the alarm.

metricAlarm_threshold :: Lens' MetricAlarm (Maybe Double)

The value to compare with the specified statistic.

metricAlarm_datapointsToAlarm :: Lens' MetricAlarm (Maybe Natural)

The number of data points that must be breaching to trigger the alarm.

metricAlarm_evaluateLowSampleCountPercentile :: Lens' MetricAlarm (Maybe Text)

Used only for alarms based on percentiles. If ignore, the alarm state does not change during periods with too few data points to be statistically significant. If evaluate or this parameter is not used, the alarm is always evaluated and possibly changes state no matter how many data points are available.

metricAlarm_comparisonOperator :: Lens' MetricAlarm (Maybe ComparisonOperator)

The arithmetic operation to use when comparing the specified statistic and threshold. The specified statistic value is used as the first operand.

metricAlarm_extendedStatistic :: Lens' MetricAlarm (Maybe Text)

The percentile statistic for the metric associated with the alarm. Specify a value between p0.0 and p100.

metricAlarm_thresholdMetricId :: Lens' MetricAlarm (Maybe Text)

In an alarm based on an anomaly detection model, this is the ID of the ANOMALY_DETECTION_BAND function used as the threshold for the alarm.

metricAlarm_alarmActions :: Lens' MetricAlarm (Maybe [Text])

The actions to execute when this alarm transitions to the ALARM state from any other state. Each action is specified as an Amazon Resource Name (ARN).

metricAlarm_unit :: Lens' MetricAlarm (Maybe StandardUnit)

The unit of the metric associated with the alarm.

metricAlarm_alarmArn :: Lens' MetricAlarm (Maybe Text)

The Amazon Resource Name (ARN) of the alarm.

metricAlarm_metricName :: Lens' MetricAlarm (Maybe Text)

The name of the metric associated with the alarm, if this is an alarm based on a single metric.

metricAlarm_stateReason :: Lens' MetricAlarm (Maybe Text)

An explanation for the alarm state, in text format.

metricAlarm_stateReasonData :: Lens' MetricAlarm (Maybe Text)

An explanation for the alarm state, in JSON format.

metricAlarm_insufficientDataActions :: Lens' MetricAlarm (Maybe [Text])

The actions to execute when this alarm transitions to the INSUFFICIENT_DATA state from any other state. Each action is specified as an Amazon Resource Name (ARN).

metricAlarm_treatMissingData :: Lens' MetricAlarm (Maybe Text)

Sets how this alarm is to handle missing data points. If this parameter is omitted, the default behavior of missing is used.

metricAlarm_metrics :: Lens' MetricAlarm (Maybe [MetricDataQuery])

An array of MetricDataQuery structures, used in an alarm based on a metric math expression. Each structure either retrieves a metric or performs a math expression. One item in the Metrics array is the math expression that the alarm watches. This expression by designated by having ReturnData set to true.

metricAlarm_alarmName :: Lens' MetricAlarm (Maybe Text)

The name of the alarm.

metricAlarm_stateValue :: Lens' MetricAlarm (Maybe StateValue)

The state value for the alarm.

metricAlarm_stateUpdatedTimestamp :: Lens' MetricAlarm (Maybe UTCTime)

The time stamp of the last update to the alarm state.

metricAlarm_oKActions :: Lens' MetricAlarm (Maybe [Text])

The actions to execute when this alarm transitions to the OK state from any other state. Each action is specified as an Amazon Resource Name (ARN).

metricAlarm_statistic :: Lens' MetricAlarm (Maybe Statistic)

The statistic for the metric associated with the alarm, other than percentile. For percentile statistics, use ExtendedStatistic.

metricAlarm_namespace :: Lens' MetricAlarm (Maybe Text)

The namespace of the metric associated with the alarm.

metricAlarm_dimensions :: Lens' MetricAlarm (Maybe [Dimension])

The dimensions for the metric associated with the alarm.

metricAlarm_evaluationPeriods :: Lens' MetricAlarm (Maybe Natural)

The number of periods over which data is compared to the specified threshold.

metricAlarm_actionsEnabled :: Lens' MetricAlarm (Maybe Bool)

Indicates whether actions should be executed during any changes to the alarm state.

metricAlarm_alarmConfigurationUpdatedTimestamp :: Lens' MetricAlarm (Maybe UTCTime)

The time stamp of the last update to the alarm configuration.

metricAlarm_period :: Lens' MetricAlarm (Maybe Natural)

The period, in seconds, over which the statistic is applied.

metricAlarm_alarmDescription :: Lens' MetricAlarm (Maybe Text)

The description of the alarm.