Network.AWS.CloudWatch.Types.MetricDataResult

Description

Documentation

data MetricDataResult

A GetMetricData call returns an array of MetricDataResult structures. Each of these structures includes the data points for that metric, along with the timestamps of those data points and other identifying information.

See: newMetricDataResult smart constructor.

Constructors

MetricDataResult'

Fields

values :: Maybe [Double]

The data points for the metric corresponding to Timestamps. The number of values always matches the number of timestamps and the timestamp for Values[x] is Timestamps[x].

id :: Maybe Text

The short name you specified to represent this metric.

timestamps :: Maybe [ISO8601]

The timestamps for the data points, formatted in Unix timestamp format. The number of timestamps always matches the number of values and the value for Timestamps[x] is Values[x].

statusCode :: Maybe StatusCode

The status of the returned data. Complete indicates that all data points in the requested time range were returned. PartialData means that an incomplete set of data points were returned. You can use the NextToken value that was returned and repeat your request to get more data points. NextToken is not returned if you are performing a math expression. InternalError indicates that an error occurred. Retry your request using NextToken, if present.

label :: Maybe Text

The human-readable label associated with the data.

messages :: Maybe [MessageData]

A list of messages with additional information about the data returned.

newMetricDataResult :: MetricDataResult

Create a value of MetricDataResult with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:values:MetricDataResult', metricDataResult_values - The data points for the metric corresponding to Timestamps. The number of values always matches the number of timestamps and the timestamp for Values[x] is Timestamps[x].

$sel:id:MetricDataResult', metricDataResult_id - The short name you specified to represent this metric.

$sel:timestamps:MetricDataResult', metricDataResult_timestamps - The timestamps for the data points, formatted in Unix timestamp format. The number of timestamps always matches the number of values and the value for Timestamps[x] is Values[x].

$sel:statusCode:MetricDataResult', metricDataResult_statusCode - The status of the returned data. Complete indicates that all data points in the requested time range were returned. PartialData means that an incomplete set of data points were returned. You can use the NextToken value that was returned and repeat your request to get more data points. NextToken is not returned if you are performing a math expression. InternalError indicates that an error occurred. Retry your request using NextToken, if present.

$sel:label:MetricDataResult', metricDataResult_label - The human-readable label associated with the data.

$sel:messages:MetricDataResult', metricDataResult_messages - A list of messages with additional information about the data returned.

metricDataResult_values :: Lens' MetricDataResult (Maybe [Double])

The data points for the metric corresponding to Timestamps. The number of values always matches the number of timestamps and the timestamp for Values[x] is Timestamps[x].

metricDataResult_id :: Lens' MetricDataResult (Maybe Text)

The short name you specified to represent this metric.

metricDataResult_timestamps :: Lens' MetricDataResult (Maybe [UTCTime])

The timestamps for the data points, formatted in Unix timestamp format. The number of timestamps always matches the number of values and the value for Timestamps[x] is Values[x].

metricDataResult_statusCode :: Lens' MetricDataResult (Maybe StatusCode)

The status of the returned data. Complete indicates that all data points in the requested time range were returned. PartialData means that an incomplete set of data points were returned. You can use the NextToken value that was returned and repeat your request to get more data points. NextToken is not returned if you are performing a math expression. InternalError indicates that an error occurred. Retry your request using NextToken, if present.

metricDataResult_label :: Lens' MetricDataResult (Maybe Text)

The human-readable label associated with the data.

metricDataResult_messages :: Lens' MetricDataResult (Maybe [MessageData])

A list of messages with additional information about the data returned.