Network.AWS.CloudWatch.Types.MetricDatum

Description

Documentation

data MetricDatum

Encapsulates the information sent to either create a metric or add new values to be aggregated into an existing metric.

See: newMetricDatum smart constructor.

Constructors

MetricDatum'

Fields

storageResolution :: Maybe Natural

Valid values are 1 and 60. Setting this to 1 specifies this metric as a high-resolution metric, so that CloudWatch stores the metric with sub-minute resolution down to one second. Setting this to 60 specifies this metric as a regular-resolution metric, which CloudWatch stores at 1-minute resolution. Currently, high resolution is available only for custom metrics. For more information about high-resolution metrics, see High-Resolution Metrics in the Amazon CloudWatch User Guide.

This field is optional, if you do not specify it the default of 60 is used.

unit :: Maybe StandardUnit

When you are using a Put operation, this defines what unit you want to use when storing the metric.

In a Get operation, this displays the unit that is used for the metric.

values :: Maybe [Double]

Array of numbers representing the values for the metric during the period. Each unique value is listed just once in this array, and the corresponding number in the Counts array specifies the number of times that value occurred during the period. You can include up to 150 unique values in each PutMetricData action that specifies a Values array.

Although the Values array accepts numbers of type Double, CloudWatch rejects values that are either too small or too large. Values must be in the range of -2^360 to 2^360. In addition, special values (for example, NaN, +Infinity, -Infinity) are not supported.

counts :: Maybe [Double]

Array of numbers that is used along with the Values array. Each number in the Count array is the number of times the corresponding value in the Values array occurred during the period.

If you omit the Counts array, the default of 1 is used as the value for each count. If you include a Counts array, it must include the same amount of values as the Values array.

statisticValues :: Maybe StatisticSet

The statistical values for the metric.

timestamp :: Maybe ISO8601

The time the metric data was received, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC.

value :: Maybe Double

The value for the metric.

Although the parameter accepts numbers of type Double, CloudWatch rejects values that are either too small or too large. Values must be in the range of -2^360 to 2^360. In addition, special values (for example, NaN, +Infinity, -Infinity) are not supported.

dimensions :: Maybe [Dimension]

The dimensions associated with the metric.

metricName :: Text

The name of the metric.

newMetricDatum

Arguments

:: Text
$sel:metricName:MetricDatum'

-> MetricDatum

Create a value of MetricDatum with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:storageResolution:MetricDatum', metricDatum_storageResolution - Valid values are 1 and 60. Setting this to 1 specifies this metric as a high-resolution metric, so that CloudWatch stores the metric with sub-minute resolution down to one second. Setting this to 60 specifies this metric as a regular-resolution metric, which CloudWatch stores at 1-minute resolution. Currently, high resolution is available only for custom metrics. For more information about high-resolution metrics, see High-Resolution Metrics in the Amazon CloudWatch User Guide.

This field is optional, if you do not specify it the default of 60 is used.

$sel:unit:MetricDatum', metricDatum_unit - When you are using a Put operation, this defines what unit you want to use when storing the metric.

In a Get operation, this displays the unit that is used for the metric.

$sel:values:MetricDatum', metricDatum_values - Array of numbers representing the values for the metric during the period. Each unique value is listed just once in this array, and the corresponding number in the Counts array specifies the number of times that value occurred during the period. You can include up to 150 unique values in each PutMetricData action that specifies a Values array.

Although the Values array accepts numbers of type Double, CloudWatch rejects values that are either too small or too large. Values must be in the range of -2^360 to 2^360. In addition, special values (for example, NaN, +Infinity, -Infinity) are not supported.

$sel:counts:MetricDatum', metricDatum_counts - Array of numbers that is used along with the Values array. Each number in the Count array is the number of times the corresponding value in the Values array occurred during the period.

If you omit the Counts array, the default of 1 is used as the value for each count. If you include a Counts array, it must include the same amount of values as the Values array.

$sel:statisticValues:MetricDatum', metricDatum_statisticValues - The statistical values for the metric.

$sel:timestamp:MetricDatum', metricDatum_timestamp - The time the metric data was received, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC.

$sel:value:MetricDatum', metricDatum_value - The value for the metric.

Although the parameter accepts numbers of type Double, CloudWatch rejects values that are either too small or too large. Values must be in the range of -2^360 to 2^360. In addition, special values (for example, NaN, +Infinity, -Infinity) are not supported.

$sel:dimensions:MetricDatum', metricDatum_dimensions - The dimensions associated with the metric.

$sel:metricName:MetricDatum', metricDatum_metricName - The name of the metric.

metricDatum_storageResolution :: Lens' MetricDatum (Maybe Natural)

Valid values are 1 and 60. Setting this to 1 specifies this metric as a high-resolution metric, so that CloudWatch stores the metric with sub-minute resolution down to one second. Setting this to 60 specifies this metric as a regular-resolution metric, which CloudWatch stores at 1-minute resolution. Currently, high resolution is available only for custom metrics. For more information about high-resolution metrics, see High-Resolution Metrics in the Amazon CloudWatch User Guide.

This field is optional, if you do not specify it the default of 60 is used.

metricDatum_unit :: Lens' MetricDatum (Maybe StandardUnit)

When you are using a Put operation, this defines what unit you want to use when storing the metric.

In a Get operation, this displays the unit that is used for the metric.

metricDatum_values :: Lens' MetricDatum (Maybe [Double])

Array of numbers representing the values for the metric during the period. Each unique value is listed just once in this array, and the corresponding number in the Counts array specifies the number of times that value occurred during the period. You can include up to 150 unique values in each PutMetricData action that specifies a Values array.

Although the Values array accepts numbers of type Double, CloudWatch rejects values that are either too small or too large. Values must be in the range of -2^360 to 2^360. In addition, special values (for example, NaN, +Infinity, -Infinity) are not supported.

metricDatum_counts :: Lens' MetricDatum (Maybe [Double])

Array of numbers that is used along with the Values array. Each number in the Count array is the number of times the corresponding value in the Values array occurred during the period.

If you omit the Counts array, the default of 1 is used as the value for each count. If you include a Counts array, it must include the same amount of values as the Values array.

metricDatum_statisticValues :: Lens' MetricDatum (Maybe StatisticSet)

The statistical values for the metric.

metricDatum_timestamp :: Lens' MetricDatum (Maybe UTCTime)

The time the metric data was received, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC.

metricDatum_value :: Lens' MetricDatum (Maybe Double)

The value for the metric.

Although the parameter accepts numbers of type Double, CloudWatch rejects values that are either too small or too large. Values must be in the range of -2^360 to 2^360. In addition, special values (for example, NaN, +Infinity, -Infinity) are not supported.

metricDatum_dimensions :: Lens' MetricDatum (Maybe [Dimension])

The dimensions associated with the metric.

metricDatum_metricName :: Lens' MetricDatum Text

The name of the metric.