Network.AWS.CloudWatch.Types.MetricStreamEntry

Description

Documentation

data MetricStreamEntry

This structure contains the configuration information about one metric stream.

See: newMetricStreamEntry smart constructor.

Constructors

MetricStreamEntry'

Fields

firehoseArn :: Maybe Text

The ARN of the Kinesis Firehose devlivery stream that is used for this metric stream.

arn :: Maybe Text

The ARN of the metric stream.

creationDate :: Maybe ISO8601

The date that the metric stream was originally created.

name :: Maybe Text

The name of the metric stream.

lastUpdateDate :: Maybe ISO8601

The date that the configuration of this metric stream was most recently updated.

state :: Maybe Text

The current state of this stream. Valid values are running and stopped.

outputFormat :: Maybe MetricStreamOutputFormat

The output format of this metric stream. Valid values are json and opentelemetry0.7.

newMetricStreamEntry :: MetricStreamEntry

Create a value of MetricStreamEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:firehoseArn:MetricStreamEntry', metricStreamEntry_firehoseArn - The ARN of the Kinesis Firehose devlivery stream that is used for this metric stream.

$sel:arn:MetricStreamEntry', metricStreamEntry_arn - The ARN of the metric stream.

$sel:creationDate:MetricStreamEntry', metricStreamEntry_creationDate - The date that the metric stream was originally created.

$sel:name:MetricStreamEntry', metricStreamEntry_name - The name of the metric stream.

$sel:lastUpdateDate:MetricStreamEntry', metricStreamEntry_lastUpdateDate - The date that the configuration of this metric stream was most recently updated.

$sel:state:MetricStreamEntry', metricStreamEntry_state - The current state of this stream. Valid values are running and stopped.

$sel:outputFormat:MetricStreamEntry', metricStreamEntry_outputFormat - The output format of this metric stream. Valid values are json and opentelemetry0.7.

metricStreamEntry_firehoseArn :: Lens' MetricStreamEntry (Maybe Text)

The ARN of the Kinesis Firehose devlivery stream that is used for this metric stream.

metricStreamEntry_arn :: Lens' MetricStreamEntry (Maybe Text)

The ARN of the metric stream.

metricStreamEntry_creationDate :: Lens' MetricStreamEntry (Maybe UTCTime)

The date that the metric stream was originally created.

metricStreamEntry_name :: Lens' MetricStreamEntry (Maybe Text)

The name of the metric stream.

metricStreamEntry_lastUpdateDate :: Lens' MetricStreamEntry (Maybe UTCTime)

The date that the configuration of this metric stream was most recently updated.

metricStreamEntry_state :: Lens' MetricStreamEntry (Maybe Text)

The current state of this stream. Valid values are running and stopped.

metricStreamEntry_outputFormat :: Lens' MetricStreamEntry (Maybe MetricStreamOutputFormat)

The output format of this metric stream. Valid values are json and opentelemetry0.7.