Network.AWS.CloudWatch.Types.PartialFailure

Description

Documentation

data PartialFailure

This array is empty if the API operation was successful for all the rules specified in the request. If the operation could not process one of the rules, the following data is returned for each of those rules.

See: newPartialFailure smart constructor.

Constructors

PartialFailure'

Fields

exceptionType :: Maybe Text

The type of error.

failureCode :: Maybe Text

The code of the error.

failureDescription :: Maybe Text

A description of the error.

failureResource :: Maybe Text

The specified rule that could not be deleted.

newPartialFailure :: PartialFailure

Create a value of PartialFailure with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:exceptionType:PartialFailure', partialFailure_exceptionType - The type of error.

$sel:failureCode:PartialFailure', partialFailure_failureCode - The code of the error.

$sel:failureDescription:PartialFailure', partialFailure_failureDescription - A description of the error.

$sel:failureResource:PartialFailure', partialFailure_failureResource - The specified rule that could not be deleted.

partialFailure_exceptionType :: Lens' PartialFailure (Maybe Text)

The type of error.

partialFailure_failureCode :: Lens' PartialFailure (Maybe Text)

The code of the error.

partialFailure_failureDescription :: Lens' PartialFailure (Maybe Text)

A description of the error.

partialFailure_failureResource :: Lens' PartialFailure (Maybe Text)

The specified rule that could not be deleted.