Network.AWS.CloudWatch.V2010_08_01

Description

Amazon CloudWatch is a monitoring service for AWS cloud resources and the applications you run on AWS. You can use Amazon CloudWatch to collect and track metrics, collect and monitor log files, and set alarms. Amazon CloudWatch can monitor AWS resources such as Amazon EC2 instances, Amazon DynamoDB tables, and Amazon RDS DB instances, as well as custom metrics generated by your applications and services, and any log files your applications generate. You can use Amazon CloudWatch to gain system-wide visibility into resource utilization, application performance, and operational health. You can use these insights to react and keep your application running smoothly.

Documentation

data DeleteAlarms

Constructors

DeleteAlarms

Fields

_dajAlarmNames :: [Text]

A list of alarms to be deleted.

dajAlarmNames :: Iso' DeleteAlarms [Text]

data DeleteAlarmsResponse

Constructors

DeleteAlarmsResponse

describeAlarmHistory :: DescribeAlarmHistory

Minimum specification for a DescribeAlarmHistory request.

data DescribeAlarmHistory

Constructors

DescribeAlarmHistory

Fields

_dahiAlarmName :: Maybe Text

The name of the alarm.

_dahiHistoryItemType :: Maybe HistoryItemType

The type of alarm histories to retrieve.

_dahiMaxRecords :: Maybe Integer

The maximum number of alarm history records to retrieve.

_dahiNextToken :: Maybe Text

The token returned by a previous call to indicate that there is more data available.

_dahiEndDate :: Maybe ISO8601

The ending date to retrieve alarm history.

_dahiStartDate :: Maybe ISO8601

The starting date to retrieve alarm history.

dahiStartDate :: Lens' DescribeAlarmHistory (Maybe ISO8601)

dahiNextToken :: Lens' DescribeAlarmHistory (Maybe Text)

dahiMaxRecords :: Lens' DescribeAlarmHistory (Maybe Integer)

dahiHistoryItemType :: Lens' DescribeAlarmHistory (Maybe HistoryItemType)

dahiEndDate :: Lens' DescribeAlarmHistory (Maybe ISO8601)

dahiAlarmName :: Lens' DescribeAlarmHistory (Maybe Text)

data DescribeAlarmHistoryResponse

Constructors

DescribeAlarmHistoryResponse

Fields

_dahoAlarmHistoryItems :: [AlarmHistoryItem]

A list of alarm histories in JSON format.

_dahoNextToken :: Maybe Text

A string that marks the start of the next batch of returned results.

dahoNextToken :: Lens' DescribeAlarmHistoryResponse (Maybe Text)

dahoAlarmHistoryItems :: Lens' DescribeAlarmHistoryResponse [AlarmHistoryItem]

describeAlarms :: DescribeAlarms

Minimum specification for a DescribeAlarms request.

data DescribeAlarms

Constructors

DescribeAlarms

Fields

_daiActionPrefix :: Maybe Text

The action name prefix.

_daiAlarmNamePrefix :: Maybe Text

The alarm name prefix. AlarmNames cannot be specified if this parameter is specified.

_daiAlarmNames :: [Text]

A list of alarm names to retrieve information for.

_daiMaxRecords :: Maybe Integer

The maximum number of alarm descriptions to retrieve.

_daiNextToken :: Maybe Text

The token returned by a previous call to indicate that there is more data available.

_daiStateValue :: Maybe StateValue

The state value to be used in matching alarms.

daiStateValue :: Lens' DescribeAlarms (Maybe StateValue)

daiNextToken :: Lens' DescribeAlarms (Maybe Text)

daiMaxRecords :: Lens' DescribeAlarms (Maybe Integer)

daiAlarmNames :: Lens' DescribeAlarms [Text]

daiAlarmNamePrefix :: Lens' DescribeAlarms (Maybe Text)

daiActionPrefix :: Lens' DescribeAlarms (Maybe Text)

data DescribeAlarmsResponse

Constructors

DescribeAlarmsResponse

Fields

_daoMetricAlarms :: [MetricAlarm]

A list of information for the specified alarms.

_daoNextToken :: Maybe Text

A string that marks the start of the next batch of returned results.

daoNextToken :: Lens' DescribeAlarmsResponse (Maybe Text)

daoMetricAlarms :: Lens' DescribeAlarmsResponse [MetricAlarm]

describeAlarmsForMetric

Arguments

:: Text
_dafmiMetricName

-> Text
_dafmiNamespace

-> DescribeAlarmsForMetric

Minimum specification for a DescribeAlarmsForMetric request.

data DescribeAlarmsForMetric

Constructors

DescribeAlarmsForMetric

Fields

_dafmiMetricName :: Text

The name of the metric.

_dafmiNamespace :: Text

The namespace of the metric.

_dafmiDimensions :: [Dimension]

The list of dimensions associated with the metric.

_dafmiPeriod :: Maybe Integer

The period in seconds over which the statistic is applied.

_dafmiUnit :: Maybe StandardUnit

The unit for the metric.

_dafmiStatistic :: Maybe Statistic

The statistic for the metric.

dafmiUnit :: Lens' DescribeAlarmsForMetric (Maybe StandardUnit)

dafmiStatistic :: Lens' DescribeAlarmsForMetric (Maybe Statistic)

dafmiPeriod :: Lens' DescribeAlarmsForMetric (Maybe Integer)

dafmiNamespace :: Lens' DescribeAlarmsForMetric Text

dafmiMetricName :: Lens' DescribeAlarmsForMetric Text

dafmiDimensions :: Lens' DescribeAlarmsForMetric [Dimension]

data DescribeAlarmsForMetricResponse

Constructors

DescribeAlarmsForMetricResponse

Fields

_dafmoMetricAlarms :: [MetricAlarm]

A list of information for each alarm with the specified metric.

dafmoMetricAlarms :: Iso' DescribeAlarmsForMetricResponse [MetricAlarm]

data DisableAlarmActions

Constructors

DisableAlarmActions

Fields

_daaiAlarmNames :: [Text]

The names of the alarms to disable actions for.

daaiAlarmNames :: Iso' DisableAlarmActions [Text]

data DisableAlarmActionsResponse

Constructors

DisableAlarmActionsResponse

data EnableAlarmActions

Constructors

EnableAlarmActions

Fields

_eaaiAlarmNames :: [Text]

The names of the alarms to enable actions for.

eaaiAlarmNames :: Iso' EnableAlarmActions [Text]

data EnableAlarmActionsResponse

Constructors

EnableAlarmActionsResponse

getMetricStatistics

Arguments

:: Text
_gmsiMetricName

-> Text
_gmsiNamespace

-> Integer
_gmsiPeriod

-> [Statistic]
_gmsiStatistics

-> ISO8601
_gmsiStartTime

-> ISO8601
_gmsiEndTime

-> GetMetricStatistics

Minimum specification for a GetMetricStatistics request.

data GetMetricStatistics

Constructors

GetMetricStatistics

Fields

_gmsiMetricName :: Text

The name of the metric.

_gmsiNamespace :: Text

The namespace of the metric.

_gmsiPeriod :: Integer

The granularity, in seconds, of the returned datapoints. Period must be at least 60 seconds and must be a multiple of 60. The default value is 60.

_gmsiStatistics :: [Statistic]

The metric statistics to return.

_gmsiStartTime :: ISO8601

The time stamp to use for determining the first datapoint to return. The value specified is inclusive; results include datapoints with the time stamp specified. The specified start time is rounded down to the nearest value. Datapoints are returned for start times up to two weeks in the past. Specified start times that are more than two weeks in the past will not return datapoints for metrics that are older than two weeks.

_gmsiEndTime :: ISO8601

The time stamp to use for determining the last datapoint to return. The value specified is exclusive; results will include datapoints up to the time stamp specified.

_gmsiDimensions :: [Dimension]

A list of dimensions describing qualities of the metric.

_gmsiUnit :: Maybe StandardUnit

The unit for the metric.

gmsiUnit :: Lens' GetMetricStatistics (Maybe StandardUnit)

gmsiStatistics :: Lens' GetMetricStatistics [Statistic]

gmsiStartTime :: Lens' GetMetricStatistics ISO8601

gmsiPeriod :: Lens' GetMetricStatistics Integer

gmsiNamespace :: Lens' GetMetricStatistics Text

gmsiMetricName :: Lens' GetMetricStatistics Text

gmsiEndTime :: Lens' GetMetricStatistics ISO8601

gmsiDimensions :: Lens' GetMetricStatistics [Dimension]

data GetMetricStatisticsResponse

Constructors

GetMetricStatisticsResponse

Fields

_gmsoDatapoints :: [Datapoint]

The datapoints for the specified metric.

_gmsoLabel :: Maybe Text

A label describing the specified metric.

gmsoLabel :: Lens' GetMetricStatisticsResponse (Maybe Text)

gmsoDatapoints :: Lens' GetMetricStatisticsResponse [Datapoint]

listMetrics :: ListMetrics

Minimum specification for a ListMetrics request.

data ListMetrics

Constructors

ListMetrics

Fields

_lmiDimensions :: [DimensionFilter]

A list of dimensions to filter against.

_lmiMetricName :: Maybe Text

The name of the metric to filter against.

_lmiNamespace :: Maybe Text

The namespace to filter against.

_lmiNextToken :: Maybe Text

The token returned by a previous call to indicate that there is more data available.

lmiNextToken :: Lens' ListMetrics (Maybe Text)

lmiNamespace :: Lens' ListMetrics (Maybe Text)

lmiMetricName :: Lens' ListMetrics (Maybe Text)

lmiDimensions :: Lens' ListMetrics [DimensionFilter]

data ListMetricsResponse

Constructors

ListMetricsResponse

Fields

_lmoMetrics :: [Metric]

A list of metrics used to generate statistics for an AWS account.

_lmoNextToken :: Maybe Text

A string that marks the start of the next batch of returned results.

lmoNextToken :: Lens' ListMetricsResponse (Maybe Text)

lmoMetrics :: Lens' ListMetricsResponse [Metric]

putMetricAlarm

Arguments

:: Text
_pmaiAlarmName

-> ComparisonOperator
_pmaiComparisonOperator

-> Integer
_pmaiEvaluationPeriods

-> Text
_pmaiMetricName

-> Text
_pmaiNamespace

-> Integer
_pmaiPeriod

-> Statistic
_pmaiStatistic

-> Double
_pmaiThreshold

-> PutMetricAlarm

Minimum specification for a PutMetricAlarm request.

data PutMetricAlarm

Constructors

PutMetricAlarm

Fields

_pmaiAlarmName :: Text

The descriptive name for the alarm. This name must be unique within the user's AWS account.

_pmaiComparisonOperator :: ComparisonOperator

The arithmetic operation to use when comparing the specified Statistic and Threshold. The specified Statistic value is used as the first operand.

_pmaiEvaluationPeriods :: Integer

The number of periods over which data is compared to the specified threshold.

_pmaiMetricName :: Text

The name for the alarm's associated metric.

_pmaiNamespace :: Text

The namespace for the alarm's associated metric.

_pmaiPeriod :: Integer

The period in seconds over which the specified statistic is applied.

_pmaiStatistic :: Statistic

The statistic to apply to the alarm's associated metric.

_pmaiThreshold :: Double

The value against which the specified statistic is compared.

_pmaiActionsEnabled :: Maybe Bool

Indicates whether or not actions should be executed during any changes to the alarm's state.

_pmaiAlarmDescription :: Maybe Text

The description for the alarm.

_pmaiDimensions :: [Dimension]

The dimensions for the alarm's associated metric.

_pmaiOKActions :: [Text]

The list of actions to execute when this alarm transitions into an OK state from any other state. Each action is specified as an Amazon Resource Number (ARN). Currently the only action supported is publishing to an Amazon SNS topic or an Amazon Auto Scaling policy.

_pmaiInsufficientDataActions :: [Text]

The list of actions to execute when this alarm transitions into an INSUFFICIENT_DATA state from any other state. Each action is specified as an Amazon Resource Number (ARN). Currently the only action supported is publishing to an Amazon SNS topic or an Amazon Auto Scaling policy.

_pmaiAlarmActions :: [Text]

The list of actions to execute when this alarm transitions into an ALARM state from any other state. Each action is specified as an Amazon Resource Number (ARN). Currently the only action supported is publishing to an Amazon SNS topic or an Amazon Auto Scaling policy.

_pmaiUnit :: Maybe StandardUnit

The unit for the alarm's associated metric.

pmaiUnit :: Lens' PutMetricAlarm (Maybe StandardUnit)

pmaiThreshold :: Lens' PutMetricAlarm Double

pmaiStatistic :: Lens' PutMetricAlarm Statistic

pmaiPeriod :: Lens' PutMetricAlarm Integer

pmaiOKActions :: Lens' PutMetricAlarm [Text]

pmaiNamespace :: Lens' PutMetricAlarm Text

pmaiMetricName :: Lens' PutMetricAlarm Text

pmaiInsufficientDataActions :: Lens' PutMetricAlarm [Text]

pmaiEvaluationPeriods :: Lens' PutMetricAlarm Integer

pmaiDimensions :: Lens' PutMetricAlarm [Dimension]

pmaiComparisonOperator :: Lens' PutMetricAlarm ComparisonOperator

pmaiAlarmName :: Lens' PutMetricAlarm Text

pmaiAlarmDescription :: Lens' PutMetricAlarm (Maybe Text)

pmaiAlarmActions :: Lens' PutMetricAlarm [Text]

pmaiActionsEnabled :: Lens' PutMetricAlarm (Maybe Bool)

data PutMetricAlarmResponse

Constructors

PutMetricAlarmResponse

data PutMetricData

Constructors

PutMetricData

Fields

_pmdiMetricData :: [MetricDatum]

A list of data describing the metric.

_pmdiNamespace :: Text

The namespace for the metric data.

pmdiNamespace :: Lens' PutMetricData Text

pmdiMetricData :: Lens' PutMetricData [MetricDatum]

data PutMetricDataResponse

Constructors

PutMetricDataResponse

setAlarmState

Arguments

:: Text
_sasiAlarmName

-> Text
_sasiStateReason

-> StateValue
_sasiStateValue

-> SetAlarmState

Minimum specification for a SetAlarmState request.

data SetAlarmState

Constructors

SetAlarmState

Fields

_sasiAlarmName :: Text

The descriptive name for the alarm. This name must be unique within the user's AWS account. The maximum length is 255 characters.

_sasiStateReason :: Text

The reason that this alarm is set to this specific state (in human-readable text format).

_sasiStateValue :: StateValue

The value of the state.

_sasiStateReasonData :: Maybe Text

The reason that this alarm is set to this specific state (in machine-readable JSON format).

sasiStateValue :: Lens' SetAlarmState StateValue

sasiStateReasonData :: Lens' SetAlarmState (Maybe Text)

sasiStateReason :: Lens' SetAlarmState Text

sasiAlarmName :: Lens' SetAlarmState Text

data SetAlarmStateResponse

Constructors

SetAlarmStateResponse

data CloudWatch

Supported version (2010-08-01) of the Amazon CloudWatch service.

xmlOptions :: Tagged a XMLOptions

data ComparisonOperator

The arithmetic operation to use when comparing the specified Statistic and Threshold. The specified Statistic value is used as the first operand.

Constructors

ComparisonOperatorGreaterThanOrEqualToThreshold
GreaterThanOrEqualToThreshold

ComparisonOperatorGreaterThanThreshold
GreaterThanThreshold

ComparisonOperatorLessThanOrEqualToThreshold
LessThanOrEqualToThreshold

ComparisonOperatorLessThanThreshold
LessThanThreshold

data HistoryItemType

The type of alarm histories to retrieve.

Constructors

HistoryItemTypeAction
Action

HistoryItemTypeConfigurationUpdate
ConfigurationUpdate

HistoryItemTypeStateUpdate
StateUpdate

data StandardUnit

The unit of the metric.

Constructors

StandardUnitBits
Bits

StandardUnitBitsSecond
Bits/Second

StandardUnitBytes
Bytes

StandardUnitBytesSecond
Bytes/Second

StandardUnitCount
Count

StandardUnitCountSecond
Count/Second

StandardUnitGigabits
Gigabits

StandardUnitGigabitsSecond
Gigabits/Second

StandardUnitGigabytes
Gigabytes

StandardUnitGigabytesSecond
Gigabytes/Second

StandardUnitKilobits
Kilobits

StandardUnitKilobitsSecond
Kilobits/Second

StandardUnitKilobytes
Kilobytes

StandardUnitKilobytesSecond
Kilobytes/Second

StandardUnitMegabits
Megabits

StandardUnitMegabitsSecond
Megabits/Second

StandardUnitMegabytes
Megabytes

StandardUnitMegabytesSecond
Megabytes/Second

StandardUnitMicroseconds
Microseconds

StandardUnitMilliseconds
Milliseconds

StandardUnitNone
None

StandardUnitPercent
Percent

StandardUnitSeconds
Seconds

StandardUnitTerabits
Terabits

StandardUnitTerabitsSecond
Terabits/Second

StandardUnitTerabytes
Terabytes

StandardUnitTerabytesSecond
Terabytes/Second

data StateValue

The state value to be used in matching alarms.

Constructors

StateValueAlarm
ALARM

StateValueInsufficientData
INSUFFICIENT_DATA

StateValueOk
OK

data Statistic

The statistic to apply to the alarm's associated metric.

Constructors

StatisticAverage
Average

StatisticMaximum
Maximum

StatisticMinimum
Minimum

StatisticSampleCount
SampleCount

StatisticSum
Sum

data AlarmHistoryItem

The AlarmHistoryItem data type contains descriptive information about the history of a specific alarm. If you call DescribeAlarmHistory, Amazon CloudWatch returns this data type as part of the DescribeAlarmHistoryResult data type.

Constructors

AlarmHistoryItem

Fields

_ahjAlarmName :: Maybe Text

The descriptive name for the alarm.

_ahjHistoryItemType :: Maybe HistoryItemType

The type of alarm history item.

_ahjHistoryData :: Maybe Text

Machine-readable data about the alarm in JSON format.

_ahjHistorySummary :: Maybe Text

A human-readable summary of the alarm history.

_ahjTimestamp :: Maybe ISO8601

The time stamp for the alarm history item.

data Datapoint

The Datapoint data type encapsulates the statistical data that Amazon CloudWatch computes from metric data.

Constructors

Datapoint

Fields

_dtSampleCount :: Maybe Double

The number of metric values that contributed to the aggregate value of this datapoint.

_dtMaximum :: Maybe Double

The maximum of the metric value used for the datapoint.

_dtAverage :: Maybe Double

The average of metric values that correspond to the datapoint.

_dtMinimum :: Maybe Double

The minimum metric value used for the datapoint.

_dtSum :: Maybe Double

The sum of metric values used for the datapoint.

_dtUnit :: Maybe StandardUnit

The standard unit used for the datapoint.

_dtTimestamp :: Maybe ISO8601

The time stamp used for the datapoint.

data Dimension

The Dimension data type further expands on the identity of a metric using a Name, Value pair.

Constructors

Dimension

Fields

_eValue :: Text

The value representing the dimension measurement.

_eName :: Text

The name of the dimension.

data DimensionFilter

The DimensionFilter data type is used to filter ListMetrics results.

Constructors

DimensionFilter

Fields

_dgValue :: Maybe Text

The value of the dimension to be matched. Specifying a Name without specifying a Value returns all values associated with that Name.

_dgName :: Text

The dimension name to be matched.

data Metric

The Metric data type contains information about a specific metric. If you call ListMetrics, Amazon CloudWatch returns information contained by this data type.

Constructors

Metric

Fields

_rMetricName :: Maybe Text

The name of the metric.

_rNamespace :: Maybe Text

The namespace of the metric.

_rDimensions :: [Dimension]

A list of dimensions associated with the metric.

data MetricAlarm

The MetricAlarm data type represents an alarm. You can use PutMetricAlarm to create or update an alarm.

Constructors

MetricAlarm

Fields

_mbAlarmName :: Maybe Text

The name of the alarm.

_mbStateUpdatedTimestamp :: Maybe ISO8601

The time stamp of the last update to the alarm's state.

_mbPeriod :: Maybe Integer

The period in seconds over which the statistic is applied.

_mbAlarmDescription :: Maybe Text

The description for the alarm.

_mbEvaluationPeriods :: Maybe Integer

The number of periods over which data is compared to the specified threshold.

_mbMetricName :: Maybe Text

The name of the alarm's metric.

_mbNamespace :: Maybe Text

The namespace of alarm's associated metric.

_mbComparisonOperator :: Maybe ComparisonOperator

The arithmetic operation to use when comparing the specified Statistic and Threshold. The specified Statistic value is used as the first operand.

_mbOKActions :: [Text]

The list of actions to execute when this alarm transitions into an OK state from any other state. Each action is specified as an Amazon Resource Number (ARN). Currently the only actions supported are publishing to an Amazon SNS topic and triggering an Auto Scaling policy.

_mbStateValue :: Maybe StateValue

The state value for the alarm.

_mbThreshold :: Maybe Double

The value against which the specified statistic is compared.

_mbAlarmConfigurationUpdatedTimestamp :: Maybe ISO8601

The time stamp of the last update to the alarm configuration.

_mbActionsEnabled :: Maybe Bool

Indicates whether actions should be executed during any changes to the alarm's state.

_mbInsufficientDataActions :: [Text]

The list of actions to execute when this alarm transitions into an INSUFFICIENT_DATA state from any other state. Each action is specified as an Amazon Resource Number (ARN). Currently the only actions supported are publishing to an Amazon SNS topic or triggering an Auto Scaling policy.

_mbStateReason :: Maybe Text

A human-readable explanation for the alarm's state.

_mbStateReasonData :: Maybe Text

An explanation for the alarm's state in machine-readable JSON format.

_mbDimensions :: [Dimension]

The list of dimensions associated with the alarm's associated metric.

_mbAlarmArn :: Maybe Text

The Amazon Resource Name (ARN) of the alarm.

_mbAlarmActions :: [Text]

The list of actions to execute when this alarm transitions into an ALARM state from any other state. Each action is specified as an Amazon Resource Number (ARN). Currently the only actions supported are publishing to an Amazon SNS topic and triggering an Auto Scaling policy.

_mbUnit :: Maybe StandardUnit

The unit of the alarm's associated metric.

_mbStatistic :: Maybe Statistic

The statistic to apply to the alarm's associated metric.

data MetricDatum

The MetricDatum data type encapsulates the information sent with PutMetricData to either create a new metric or add new values to be aggregated into an existing metric.

Constructors

MetricDatum

Fields

_meMetricName :: Text

The name of the metric.

_meValue :: Maybe Double

The value for the metric. Although the Value parameter accepts numbers of type Double, Amazon CloudWatch truncates values with very large exponents. Values with base-10 exponents greater than 126 (1 x 10^126) are truncated. Likewise, values with base-10 exponents less than -130 (1 x 10^-130) are also truncated.

_meDimensions :: [Dimension]

A list of dimensions associated with the metric.

_meUnit :: Maybe StandardUnit

The unit of the metric.

_meTimestamp :: Maybe ISO8601

The time stamp used for the metric. If not specified, the default value is set to the time the metric data was received.

_meStatisticValues :: Maybe StatisticSet

A set of statistical values describing the metric.

data StatisticSet

A set of statistical values describing the metric.

Constructors

StatisticSet

Fields

_ssSampleCount :: Double

The number of samples used for the statistic set.

_ssMaximum :: Double

The maximum value of the sample set.

_ssMinimum :: Double

The minimum value of the sample set.

_ssSum :: Double

The sum of values for the sample set.

ahjTimestamp :: Lens' AlarmHistoryItem (Maybe ISO8601)

ahjHistorySummary :: Lens' AlarmHistoryItem (Maybe Text)

ahjHistoryItemType :: Lens' AlarmHistoryItem (Maybe HistoryItemType)

ahjHistoryData :: Lens' AlarmHistoryItem (Maybe Text)

ahjAlarmName :: Lens' AlarmHistoryItem (Maybe Text)

dtUnit :: Lens' Datapoint (Maybe StandardUnit)

dtTimestamp :: Lens' Datapoint (Maybe ISO8601)

dtSum :: Lens' Datapoint (Maybe Double)

dtSampleCount :: Lens' Datapoint (Maybe Double)

dtMinimum :: Lens' Datapoint (Maybe Double)

dtMaximum :: Lens' Datapoint (Maybe Double)

dtAverage :: Lens' Datapoint (Maybe Double)

eValue :: Lens' Dimension Text

eName :: Lens' Dimension Text

dgValue :: Lens' DimensionFilter (Maybe Text)

dgName :: Lens' DimensionFilter Text

rNamespace :: Lens' Metric (Maybe Text)

rMetricName :: Lens' Metric (Maybe Text)

rDimensions :: Lens' Metric [Dimension]

mbUnit :: Lens' MetricAlarm (Maybe StandardUnit)

mbThreshold :: Lens' MetricAlarm (Maybe Double)

mbStatistic :: Lens' MetricAlarm (Maybe Statistic)

mbStateValue :: Lens' MetricAlarm (Maybe StateValue)

mbStateUpdatedTimestamp :: Lens' MetricAlarm (Maybe ISO8601)

mbStateReasonData :: Lens' MetricAlarm (Maybe Text)

mbStateReason :: Lens' MetricAlarm (Maybe Text)

mbPeriod :: Lens' MetricAlarm (Maybe Integer)

mbOKActions :: Lens' MetricAlarm [Text]

mbNamespace :: Lens' MetricAlarm (Maybe Text)

mbMetricName :: Lens' MetricAlarm (Maybe Text)

mbInsufficientDataActions :: Lens' MetricAlarm [Text]

mbEvaluationPeriods :: Lens' MetricAlarm (Maybe Integer)

mbDimensions :: Lens' MetricAlarm [Dimension]

mbComparisonOperator :: Lens' MetricAlarm (Maybe ComparisonOperator)

mbAlarmName :: Lens' MetricAlarm (Maybe Text)

mbAlarmDescription :: Lens' MetricAlarm (Maybe Text)

mbAlarmConfigurationUpdatedTimestamp :: Lens' MetricAlarm (Maybe ISO8601)

mbAlarmArn :: Lens' MetricAlarm (Maybe Text)

mbAlarmActions :: Lens' MetricAlarm [Text]

mbActionsEnabled :: Lens' MetricAlarm (Maybe Bool)

meValue :: Lens' MetricDatum (Maybe Double)

meUnit :: Lens' MetricDatum (Maybe StandardUnit)

meTimestamp :: Lens' MetricDatum (Maybe ISO8601)

meStatisticValues :: Lens' MetricDatum (Maybe StatisticSet)

meMetricName :: Lens' MetricDatum Text

meDimensions :: Lens' MetricDatum [Dimension]

ssSum :: Lens' StatisticSet Double

ssSampleCount :: Lens' StatisticSet Double

ssMinimum :: Lens' StatisticSet Double

ssMaximum :: Lens' StatisticSet Double