Network.AWS.CloudWatch.V2010_08_01.DescribeAlarmHistory

Description

Retrieves history for the specified alarm. Filter alarms by date range or item type. If an alarm name is not specified, Amazon CloudWatch returns histories for all of the owner's alarms. Amazon CloudWatch retains the history of an alarm for two weeks, whether or not you delete the alarm.

Documentation

describeAlarmHistory :: DescribeAlarmHistory

Minimum specification for a DescribeAlarmHistory request.

data DescribeAlarmHistory

Constructors

DescribeAlarmHistory

Fields

_dahiAlarmName :: Maybe Text

The name of the alarm.

_dahiHistoryItemType :: Maybe HistoryItemType

The type of alarm histories to retrieve.

_dahiMaxRecords :: Maybe Integer

The maximum number of alarm history records to retrieve.

_dahiNextToken :: Maybe Text

The token returned by a previous call to indicate that there is more data available.

_dahiEndDate :: Maybe ISO8601

The ending date to retrieve alarm history.

_dahiStartDate :: Maybe ISO8601

The starting date to retrieve alarm history.

dahiStartDate :: Lens' DescribeAlarmHistory (Maybe ISO8601)

dahiNextToken :: Lens' DescribeAlarmHistory (Maybe Text)

dahiMaxRecords :: Lens' DescribeAlarmHistory (Maybe Integer)

dahiHistoryItemType :: Lens' DescribeAlarmHistory (Maybe HistoryItemType)

dahiEndDate :: Lens' DescribeAlarmHistory (Maybe ISO8601)

dahiAlarmName :: Lens' DescribeAlarmHistory (Maybe Text)

data DescribeAlarmHistoryResponse

Constructors

DescribeAlarmHistoryResponse

Fields

_dahoAlarmHistoryItems :: [AlarmHistoryItem]

A list of alarm histories in JSON format.

_dahoNextToken :: Maybe Text

A string that marks the start of the next batch of returned results.

dahoNextToken :: Lens' DescribeAlarmHistoryResponse (Maybe Text)

dahoAlarmHistoryItems :: Lens' DescribeAlarmHistoryResponse [AlarmHistoryItem]