Network.AWS.CloudWatch.V2010_08_01.DisableAlarmActions

Description

Disables actions for the specified alarms. When an alarm's actions are disabled the alarm's state may change, but none of the alarm's actions will execute.

Documentation

data DisableAlarmActions

Constructors

DisableAlarmActions

Fields

_daaiAlarmNames :: [Text]

The names of the alarms to disable actions for.

daaiAlarmNames :: Iso' DisableAlarmActions [Text]

data DisableAlarmActionsResponse

Constructors

DisableAlarmActionsResponse