Network.AWS.CloudWatch.V2010_08_01.GetMetricStatistics

Description

Gets statistics for the specified metric. The maximum number of data points returned from a single GetMetricStatistics request is 1,440. If a request is made that generates more than 1,440 data points, Amazon CloudWatch returns an error. In such a case, alter the request by narrowing the specified time range or increasing the specified period. Alternatively, make multiple requests across adjacent time ranges. Amazon CloudWatch aggregates data points based on the length of the period that you specify. For example, if you request statistics with a one-minute granularity, Amazon CloudWatch aggregates data points with time stamps that fall within the same one-minute period. In such a case, the data points queried can greatly outnumber the data points returned. The maximum number of data points that can be queried is 50,850; whereas the maximum number of data points returned is 1,440. The following examples show various statistics allowed by the data point query maximum of 50,850 when you call GetMetricStatistics on Amazon EC2 instances with detailed (one-minute) monitoring enabled: Statistics for up to 400 instances for a span of one hour Statistics for up to 35 instances over a span of 24 hours Statistics for up to 2 instances over a span of 2 weeks.

Documentation

getMetricStatistics

Arguments

:: Text
_gmsiMetricName

-> Text
_gmsiNamespace

-> Integer
_gmsiPeriod

-> [Statistic]
_gmsiStatistics

-> ISO8601
_gmsiStartTime

-> ISO8601
_gmsiEndTime

-> GetMetricStatistics

Minimum specification for a GetMetricStatistics request.

data GetMetricStatistics

Constructors

GetMetricStatistics

Fields

_gmsiMetricName :: Text

The name of the metric.

_gmsiNamespace :: Text

The namespace of the metric.

_gmsiPeriod :: Integer

The granularity, in seconds, of the returned datapoints. Period must be at least 60 seconds and must be a multiple of 60. The default value is 60.

_gmsiStatistics :: [Statistic]

The metric statistics to return.

_gmsiStartTime :: ISO8601

The time stamp to use for determining the first datapoint to return. The value specified is inclusive; results include datapoints with the time stamp specified. The specified start time is rounded down to the nearest value. Datapoints are returned for start times up to two weeks in the past. Specified start times that are more than two weeks in the past will not return datapoints for metrics that are older than two weeks.

_gmsiEndTime :: ISO8601

The time stamp to use for determining the last datapoint to return. The value specified is exclusive; results will include datapoints up to the time stamp specified.

_gmsiDimensions :: [Dimension]

A list of dimensions describing qualities of the metric.

_gmsiUnit :: Maybe StandardUnit

The unit for the metric.

gmsiUnit :: Lens' GetMetricStatistics (Maybe StandardUnit)

gmsiStatistics :: Lens' GetMetricStatistics [Statistic]

gmsiStartTime :: Lens' GetMetricStatistics ISO8601

gmsiPeriod :: Lens' GetMetricStatistics Integer

gmsiNamespace :: Lens' GetMetricStatistics Text

gmsiMetricName :: Lens' GetMetricStatistics Text

gmsiEndTime :: Lens' GetMetricStatistics ISO8601

gmsiDimensions :: Lens' GetMetricStatistics [Dimension]

data GetMetricStatisticsResponse

Constructors

GetMetricStatisticsResponse

Fields

_gmsoDatapoints :: [Datapoint]

The datapoints for the specified metric.

_gmsoLabel :: Maybe Text

A label describing the specified metric.

gmsoLabel :: Lens' GetMetricStatisticsResponse (Maybe Text)

gmsoDatapoints :: Lens' GetMetricStatisticsResponse [Datapoint]