Network.AWS.CloudWatch.V2010_08_01.PutMetricData

Description

Publishes metric data points to Amazon CloudWatch. Amazon Cloudwatch associates the data points with the specified metric. If the specified metric does not exist, Amazon CloudWatch creates the metric. If you create a metric with the PutMetricData action, allow up to fifteen minutes for the metric to appear in calls to the ListMetrics action. The size of a PutMetricData request is limited to 8 KB for HTTP GET requests and 40 KB for HTTP POST requests. Although the Value parameter accepts numbers of type Double, Amazon CloudWatch truncates values with very large exponents. Values with base-10 exponents greater than 126 (1 x 10^126) are truncated. Likewise, values with base-10 exponents less than -130 (1 x 10^-130) are also truncated.

Documentation

data PutMetricData

Constructors

PutMetricData

Fields

_pmdiMetricData :: [MetricDatum]

A list of data describing the metric.

_pmdiNamespace :: Text

The namespace for the metric data.

pmdiNamespace :: Lens' PutMetricData Text

pmdiMetricData :: Lens' PutMetricData [MetricDatum]

data PutMetricDataResponse

Constructors

PutMetricDataResponse