Network.AWS.CloudWatch.V2010_08_01.SetAlarmState

Description

Temporarily sets the state of an alarm. When the updated StateValue differs from the previous value, the action configured for the appropriate state is invoked. This is not a permanent change. The next periodic alarm check (in about a minute) will set the alarm to its actual state.

Documentation

setAlarmState

Arguments

:: Text
_sasiAlarmName

-> Text
_sasiStateReason

-> StateValue
_sasiStateValue

-> SetAlarmState

Minimum specification for a SetAlarmState request.

data SetAlarmState

Constructors

SetAlarmState

Fields

_sasiAlarmName :: Text

The descriptive name for the alarm. This name must be unique within the user's AWS account. The maximum length is 255 characters.

_sasiStateReason :: Text

The reason that this alarm is set to this specific state (in human-readable text format).

_sasiStateValue :: StateValue

The value of the state.

_sasiStateReasonData :: Maybe Text

The reason that this alarm is set to this specific state (in machine-readable JSON format).

sasiStateValue :: Lens' SetAlarmState StateValue

sasiStateReasonData :: Lens' SetAlarmState (Maybe Text)

sasiStateReason :: Lens' SetAlarmState Text

sasiAlarmName :: Lens' SetAlarmState Text

data SetAlarmStateResponse

Constructors

SetAlarmStateResponse