Network.AWS.CloudWatchEvents.CreateApiDestination

Description

Creates an API destination, which is an HTTP invocation endpoint configured as a target for events.

Creating a Request

data CreateApiDestination

See: newCreateApiDestination smart constructor.

Constructors

CreateApiDestination'

Fields

description :: Maybe Text

A description for the API destination to create.

invocationRateLimitPerSecond :: Maybe Natural

The maximum number of requests per second to send to the HTTP invocation endpoint.

name :: Text

The name for the API destination to create.

connectionArn :: Text

The ARN of the connection to use for the API destination. The destination endpoint must support the authorization type specified for the connection.

invocationEndpoint :: Text

The URL to the HTTP invocation endpoint for the API destination.

httpMethod :: ApiDestinationHttpMethod

The method to use for the request to the HTTP invocation endpoint.

newCreateApiDestination

Arguments

:: Text
$sel:name:CreateApiDestination'

-> Text
$sel:connectionArn:CreateApiDestination'

-> Text
$sel:invocationEndpoint:CreateApiDestination'

-> ApiDestinationHttpMethod
$sel:httpMethod:CreateApiDestination'

-> CreateApiDestination

Create a value of CreateApiDestination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:CreateApiDestination', createApiDestination_description - A description for the API destination to create.

$sel:invocationRateLimitPerSecond:CreateApiDestination', createApiDestination_invocationRateLimitPerSecond - The maximum number of requests per second to send to the HTTP invocation endpoint.

$sel:name:CreateApiDestination', createApiDestination_name - The name for the API destination to create.

$sel:connectionArn:CreateApiDestination', createApiDestination_connectionArn - The ARN of the connection to use for the API destination. The destination endpoint must support the authorization type specified for the connection.

$sel:invocationEndpoint:CreateApiDestination', createApiDestination_invocationEndpoint - The URL to the HTTP invocation endpoint for the API destination.

$sel:httpMethod:CreateApiDestination', createApiDestination_httpMethod - The method to use for the request to the HTTP invocation endpoint.

Request Lenses

createApiDestination_description :: Lens' CreateApiDestination (Maybe Text)

A description for the API destination to create.

createApiDestination_invocationRateLimitPerSecond :: Lens' CreateApiDestination (Maybe Natural)

The maximum number of requests per second to send to the HTTP invocation endpoint.

createApiDestination_name :: Lens' CreateApiDestination Text

The name for the API destination to create.

createApiDestination_connectionArn :: Lens' CreateApiDestination Text

The ARN of the connection to use for the API destination. The destination endpoint must support the authorization type specified for the connection.

createApiDestination_invocationEndpoint :: Lens' CreateApiDestination Text

The URL to the HTTP invocation endpoint for the API destination.

createApiDestination_httpMethod :: Lens' CreateApiDestination ApiDestinationHttpMethod

The method to use for the request to the HTTP invocation endpoint.

Destructuring the Response

data CreateApiDestinationResponse

See: newCreateApiDestinationResponse smart constructor.

Constructors

CreateApiDestinationResponse'

Fields

creationTime :: Maybe POSIX

A time stamp indicating the time that the API destination was created.

apiDestinationArn :: Maybe Text

The ARN of the API destination that was created by the request.

apiDestinationState :: Maybe ApiDestinationState

The state of the API destination that was created by the request.

lastModifiedTime :: Maybe POSIX

A time stamp indicating the time that the API destination was last modified.

httpStatus :: Int

The response's http status code.

newCreateApiDestinationResponse

Arguments

:: Int
$sel:httpStatus:CreateApiDestinationResponse'

-> CreateApiDestinationResponse

Create a value of CreateApiDestinationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:CreateApiDestinationResponse', createApiDestinationResponse_creationTime - A time stamp indicating the time that the API destination was created.

$sel:apiDestinationArn:CreateApiDestinationResponse', createApiDestinationResponse_apiDestinationArn - The ARN of the API destination that was created by the request.

$sel:apiDestinationState:CreateApiDestinationResponse', createApiDestinationResponse_apiDestinationState - The state of the API destination that was created by the request.

$sel:lastModifiedTime:CreateApiDestinationResponse', createApiDestinationResponse_lastModifiedTime - A time stamp indicating the time that the API destination was last modified.

$sel:httpStatus:CreateApiDestinationResponse', createApiDestinationResponse_httpStatus - The response's http status code.

Response Lenses

createApiDestinationResponse_creationTime :: Lens' CreateApiDestinationResponse (Maybe UTCTime)

A time stamp indicating the time that the API destination was created.

createApiDestinationResponse_apiDestinationArn :: Lens' CreateApiDestinationResponse (Maybe Text)

The ARN of the API destination that was created by the request.

createApiDestinationResponse_apiDestinationState :: Lens' CreateApiDestinationResponse (Maybe ApiDestinationState)

The state of the API destination that was created by the request.

createApiDestinationResponse_lastModifiedTime :: Lens' CreateApiDestinationResponse (Maybe UTCTime)

A time stamp indicating the time that the API destination was last modified.

createApiDestinationResponse_httpStatus :: Lens' CreateApiDestinationResponse Int

The response's http status code.