Network.AWS.CloudWatchEvents.CreateArchive

Description

Creates an archive of events with the specified settings. When you create an archive, incoming events might not immediately start being sent to the archive. Allow a short period of time for changes to take effect. If you do not specify a pattern to filter events sent to the archive, all events are sent to the archive except replayed events. Replayed events are not sent to an archive.

Creating a Request

data CreateArchive

See: newCreateArchive smart constructor.

Constructors

CreateArchive'

Fields

eventPattern :: Maybe Text

An event pattern to use to filter events sent to the archive.

description :: Maybe Text

A description for the archive.

retentionDays :: Maybe Natural

The number of days to retain events for. Default value is 0. If set to 0, events are retained indefinitely

archiveName :: Text

The name for the archive to create.

eventSourceArn :: Text

The ARN of the event bus that sends events to the archive.

newCreateArchive

Arguments

:: Text
$sel:archiveName:CreateArchive'

-> Text
$sel:eventSourceArn:CreateArchive'

-> CreateArchive

Create a value of CreateArchive with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventPattern:CreateArchive', createArchive_eventPattern - An event pattern to use to filter events sent to the archive.

$sel:description:CreateArchive', createArchive_description - A description for the archive.

$sel:retentionDays:CreateArchive', createArchive_retentionDays - The number of days to retain events for. Default value is 0. If set to 0, events are retained indefinitely

$sel:archiveName:CreateArchive', createArchive_archiveName - The name for the archive to create.

$sel:eventSourceArn:CreateArchive', createArchive_eventSourceArn - The ARN of the event bus that sends events to the archive.

Request Lenses

createArchive_eventPattern :: Lens' CreateArchive (Maybe Text)

An event pattern to use to filter events sent to the archive.

createArchive_description :: Lens' CreateArchive (Maybe Text)

A description for the archive.

createArchive_retentionDays :: Lens' CreateArchive (Maybe Natural)

The number of days to retain events for. Default value is 0. If set to 0, events are retained indefinitely

createArchive_archiveName :: Lens' CreateArchive Text

The name for the archive to create.

createArchive_eventSourceArn :: Lens' CreateArchive Text

The ARN of the event bus that sends events to the archive.

Destructuring the Response

data CreateArchiveResponse

See: newCreateArchiveResponse smart constructor.

Constructors

CreateArchiveResponse'

Fields

creationTime :: Maybe POSIX

The time at which the archive was created.

stateReason :: Maybe Text

The reason that the archive is in the state.

archiveArn :: Maybe Text

The ARN of the archive that was created.

state :: Maybe ArchiveState

The state of the archive that was created.

httpStatus :: Int

The response's http status code.

newCreateArchiveResponse

Arguments

:: Int
$sel:httpStatus:CreateArchiveResponse'

-> CreateArchiveResponse

Create a value of CreateArchiveResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:CreateArchiveResponse', createArchiveResponse_creationTime - The time at which the archive was created.

$sel:stateReason:CreateArchiveResponse', createArchiveResponse_stateReason - The reason that the archive is in the state.

$sel:archiveArn:CreateArchiveResponse', createArchiveResponse_archiveArn - The ARN of the archive that was created.

$sel:state:CreateArchiveResponse', createArchiveResponse_state - The state of the archive that was created.

$sel:httpStatus:CreateArchiveResponse', createArchiveResponse_httpStatus - The response's http status code.

Response Lenses

createArchiveResponse_creationTime :: Lens' CreateArchiveResponse (Maybe UTCTime)

The time at which the archive was created.

createArchiveResponse_stateReason :: Lens' CreateArchiveResponse (Maybe Text)

The reason that the archive is in the state.

createArchiveResponse_archiveArn :: Lens' CreateArchiveResponse (Maybe Text)

The ARN of the archive that was created.

createArchiveResponse_state :: Lens' CreateArchiveResponse (Maybe ArchiveState)

The state of the archive that was created.

createArchiveResponse_httpStatus :: Lens' CreateArchiveResponse Int

The response's http status code.