Network.AWS.CloudWatchEvents.CreateConnection

Description

Creates a connection. A connection defines the authorization type and credentials to use for authorization with an API destination HTTP endpoint.

Creating a Request

data CreateConnection

See: newCreateConnection smart constructor.

Constructors

CreateConnection'

Fields

description :: Maybe Text

A description for the connection to create.

name :: Text

The name for the connection to create.

authorizationType :: ConnectionAuthorizationType

The type of authorization to use for the connection.

authParameters :: CreateConnectionAuthRequestParameters

A CreateConnectionAuthRequestParameters object that contains the authorization parameters to use to authorize with the endpoint.

newCreateConnection

Arguments

:: Text
$sel:name:CreateConnection'

-> ConnectionAuthorizationType
$sel:authorizationType:CreateConnection'

-> CreateConnectionAuthRequestParameters
$sel:authParameters:CreateConnection'

-> CreateConnection

Create a value of CreateConnection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:CreateConnection', createConnection_description - A description for the connection to create.

$sel:name:CreateConnection', createConnection_name - The name for the connection to create.

$sel:authorizationType:CreateConnection', createConnection_authorizationType - The type of authorization to use for the connection.

$sel:authParameters:CreateConnection', createConnection_authParameters - A CreateConnectionAuthRequestParameters object that contains the authorization parameters to use to authorize with the endpoint.

Request Lenses

createConnection_description :: Lens' CreateConnection (Maybe Text)

A description for the connection to create.

createConnection_name :: Lens' CreateConnection Text

The name for the connection to create.

createConnection_authorizationType :: Lens' CreateConnection ConnectionAuthorizationType

The type of authorization to use for the connection.

createConnection_authParameters :: Lens' CreateConnection CreateConnectionAuthRequestParameters

A CreateConnectionAuthRequestParameters object that contains the authorization parameters to use to authorize with the endpoint.

Destructuring the Response

data CreateConnectionResponse

See: newCreateConnectionResponse smart constructor.

Constructors

CreateConnectionResponse'

Fields

creationTime :: Maybe POSIX

A time stamp for the time that the connection was created.

lastModifiedTime :: Maybe POSIX

A time stamp for the time that the connection was last updated.

connectionArn :: Maybe Text

The ARN of the connection that was created by the request.

connectionState :: Maybe ConnectionState

The state of the connection that was created by the request.

httpStatus :: Int

The response's http status code.

newCreateConnectionResponse

Arguments

:: Int
$sel:httpStatus:CreateConnectionResponse'

-> CreateConnectionResponse

Create a value of CreateConnectionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:CreateConnectionResponse', createConnectionResponse_creationTime - A time stamp for the time that the connection was created.

$sel:lastModifiedTime:CreateConnectionResponse', createConnectionResponse_lastModifiedTime - A time stamp for the time that the connection was last updated.

$sel:connectionArn:CreateConnectionResponse', createConnectionResponse_connectionArn - The ARN of the connection that was created by the request.

$sel:connectionState:CreateConnectionResponse', createConnectionResponse_connectionState - The state of the connection that was created by the request.

$sel:httpStatus:CreateConnectionResponse', createConnectionResponse_httpStatus - The response's http status code.

Response Lenses

createConnectionResponse_creationTime :: Lens' CreateConnectionResponse (Maybe UTCTime)

A time stamp for the time that the connection was created.

createConnectionResponse_lastModifiedTime :: Lens' CreateConnectionResponse (Maybe UTCTime)

A time stamp for the time that the connection was last updated.

createConnectionResponse_connectionArn :: Lens' CreateConnectionResponse (Maybe Text)

The ARN of the connection that was created by the request.

createConnectionResponse_connectionState :: Lens' CreateConnectionResponse (Maybe ConnectionState)

The state of the connection that was created by the request.

createConnectionResponse_httpStatus :: Lens' CreateConnectionResponse Int

The response's http status code.