Network.AWS.CloudWatchEvents.CreatePartnerEventSource

Description

Called by an SaaS partner to create a partner event source. This operation is not used by Amazon Web Services customers.

Each partner event source can be used by one Amazon Web Services account to create a matching partner event bus in that Amazon Web Services account. A SaaS partner must create one partner event source for each Amazon Web Services account that wants to receive those event types.

A partner event source creates events based on resources within the SaaS partner's service or application.

An Amazon Web Services account that creates a partner event bus that matches the partner event source can use that event bus to receive events from the partner, and then process them using Amazon Web Services Events rules and targets.

Partner event source names follow this format:

 partner_name/event_namespace/event_name

partner_name is determined during partner registration and identifies the partner to Amazon Web Services customers. event_namespace is determined by the partner and is a way for the partner to categorize their events. event_name is determined by the partner, and should uniquely identify an event-generating resource within the partner system. The combination of event_namespace and event_name should help Amazon Web Services customers decide whether to create an event bus to receive these events.

Creating a Request

data CreatePartnerEventSource

See: newCreatePartnerEventSource smart constructor.

Constructors

CreatePartnerEventSource'

Fields

name :: Text

The name of the partner event source. This name must be unique and must be in the format partner_name/event_namespace/event_name . The Amazon Web Services account that wants to use this partner event source must create a partner event bus with a name that matches the name of the partner event source.

account :: Text

The Amazon Web Services account ID that is permitted to create a matching partner event bus for this partner event source.

newCreatePartnerEventSource

Arguments

:: Text
$sel:name:CreatePartnerEventSource'

-> Text
$sel:account:CreatePartnerEventSource'

-> CreatePartnerEventSource

Create a value of CreatePartnerEventSource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:CreatePartnerEventSource', createPartnerEventSource_name - The name of the partner event source. This name must be unique and must be in the format partner_name/event_namespace/event_name . The Amazon Web Services account that wants to use this partner event source must create a partner event bus with a name that matches the name of the partner event source.

$sel:account:CreatePartnerEventSource', createPartnerEventSource_account - The Amazon Web Services account ID that is permitted to create a matching partner event bus for this partner event source.

Request Lenses

createPartnerEventSource_name :: Lens' CreatePartnerEventSource Text

The name of the partner event source. This name must be unique and must be in the format partner_name/event_namespace/event_name . The Amazon Web Services account that wants to use this partner event source must create a partner event bus with a name that matches the name of the partner event source.

createPartnerEventSource_account :: Lens' CreatePartnerEventSource Text

The Amazon Web Services account ID that is permitted to create a matching partner event bus for this partner event source.

Destructuring the Response

data CreatePartnerEventSourceResponse

See: newCreatePartnerEventSourceResponse smart constructor.

Constructors

CreatePartnerEventSourceResponse'

Fields

eventSourceArn :: Maybe Text

The ARN of the partner event source.

httpStatus :: Int

The response's http status code.

newCreatePartnerEventSourceResponse

Arguments

:: Int
$sel:httpStatus:CreatePartnerEventSourceResponse'

-> CreatePartnerEventSourceResponse

Create a value of CreatePartnerEventSourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventSourceArn:CreatePartnerEventSourceResponse', createPartnerEventSourceResponse_eventSourceArn - The ARN of the partner event source.

$sel:httpStatus:CreatePartnerEventSourceResponse', createPartnerEventSourceResponse_httpStatus - The response's http status code.

Response Lenses

createPartnerEventSourceResponse_eventSourceArn :: Lens' CreatePartnerEventSourceResponse (Maybe Text)

The ARN of the partner event source.

createPartnerEventSourceResponse_httpStatus :: Lens' CreatePartnerEventSourceResponse Int

The response's http status code.