Network.AWS.CloudWatchEvents.DescribeArchive

Description

Retrieves details about an archive.

Creating a Request

data DescribeArchive

See: newDescribeArchive smart constructor.

Constructors

DescribeArchive'

Fields

archiveName :: Text

The name of the archive to retrieve.

newDescribeArchive

Arguments

:: Text
$sel:archiveName:DescribeArchive'

-> DescribeArchive

Create a value of DescribeArchive with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:archiveName:DescribeArchive', describeArchive_archiveName - The name of the archive to retrieve.

Request Lenses

describeArchive_archiveName :: Lens' DescribeArchive Text

The name of the archive to retrieve.

Destructuring the Response

data DescribeArchiveResponse

See: newDescribeArchiveResponse smart constructor.

Constructors

DescribeArchiveResponse'

Fields

eventCount :: Maybe Integer

The number of events in the archive.

creationTime :: Maybe POSIX

The time at which the archive was created.

eventSourceArn :: Maybe Text

The ARN of the event source associated with the archive.

eventPattern :: Maybe Text

The event pattern used to filter events sent to the archive.

stateReason :: Maybe Text

The reason that the archive is in the state.

archiveName :: Maybe Text

The name of the archive.

archiveArn :: Maybe Text

The ARN of the archive.

state :: Maybe ArchiveState

The state of the archive.

sizeBytes :: Maybe Integer

The size of the archive in bytes.

description :: Maybe Text

The description of the archive.

retentionDays :: Maybe Natural

The number of days to retain events for in the archive.

httpStatus :: Int

The response's http status code.

newDescribeArchiveResponse

Arguments

:: Int
$sel:httpStatus:DescribeArchiveResponse'

-> DescribeArchiveResponse

Create a value of DescribeArchiveResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventCount:DescribeArchiveResponse', describeArchiveResponse_eventCount - The number of events in the archive.

$sel:creationTime:DescribeArchiveResponse', describeArchiveResponse_creationTime - The time at which the archive was created.

$sel:eventSourceArn:DescribeArchiveResponse', describeArchiveResponse_eventSourceArn - The ARN of the event source associated with the archive.

$sel:eventPattern:DescribeArchiveResponse', describeArchiveResponse_eventPattern - The event pattern used to filter events sent to the archive.

$sel:stateReason:DescribeArchiveResponse', describeArchiveResponse_stateReason - The reason that the archive is in the state.

$sel:archiveName:DescribeArchive', describeArchiveResponse_archiveName - The name of the archive.

$sel:archiveArn:DescribeArchiveResponse', describeArchiveResponse_archiveArn - The ARN of the archive.

$sel:state:DescribeArchiveResponse', describeArchiveResponse_state - The state of the archive.

$sel:sizeBytes:DescribeArchiveResponse', describeArchiveResponse_sizeBytes - The size of the archive in bytes.

$sel:description:DescribeArchiveResponse', describeArchiveResponse_description - The description of the archive.

$sel:retentionDays:DescribeArchiveResponse', describeArchiveResponse_retentionDays - The number of days to retain events for in the archive.

$sel:httpStatus:DescribeArchiveResponse', describeArchiveResponse_httpStatus - The response's http status code.

Response Lenses

describeArchiveResponse_eventCount :: Lens' DescribeArchiveResponse (Maybe Integer)

The number of events in the archive.

describeArchiveResponse_creationTime :: Lens' DescribeArchiveResponse (Maybe UTCTime)

The time at which the archive was created.

describeArchiveResponse_eventSourceArn :: Lens' DescribeArchiveResponse (Maybe Text)

The ARN of the event source associated with the archive.

describeArchiveResponse_eventPattern :: Lens' DescribeArchiveResponse (Maybe Text)

The event pattern used to filter events sent to the archive.

describeArchiveResponse_stateReason :: Lens' DescribeArchiveResponse (Maybe Text)

The reason that the archive is in the state.

describeArchiveResponse_archiveName :: Lens' DescribeArchiveResponse (Maybe Text)

The name of the archive.

describeArchiveResponse_archiveArn :: Lens' DescribeArchiveResponse (Maybe Text)

The ARN of the archive.

describeArchiveResponse_state :: Lens' DescribeArchiveResponse (Maybe ArchiveState)

The state of the archive.

describeArchiveResponse_sizeBytes :: Lens' DescribeArchiveResponse (Maybe Integer)

The size of the archive in bytes.

describeArchiveResponse_description :: Lens' DescribeArchiveResponse (Maybe Text)

The description of the archive.

describeArchiveResponse_retentionDays :: Lens' DescribeArchiveResponse (Maybe Natural)

The number of days to retain events for in the archive.

describeArchiveResponse_httpStatus :: Lens' DescribeArchiveResponse Int

The response's http status code.