Network.AWS.CloudWatchEvents.DescribeConnection

Description

Retrieves details about a connection.

Creating a Request

data DescribeConnection

See: newDescribeConnection smart constructor.

Constructors

DescribeConnection'

Fields

name :: Text

The name of the connection to retrieve.

newDescribeConnection

Arguments

:: Text
$sel:name:DescribeConnection'

-> DescribeConnection

Create a value of DescribeConnection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:DescribeConnection', describeConnection_name - The name of the connection to retrieve.

Request Lenses

describeConnection_name :: Lens' DescribeConnection Text

The name of the connection to retrieve.

Destructuring the Response

data DescribeConnectionResponse

See: newDescribeConnectionResponse smart constructor.

Constructors

DescribeConnectionResponse'

Fields

creationTime :: Maybe POSIX

A time stamp for the time that the connection was created.

connectionState :: Maybe ConnectionState

The state of the connection retrieved.

stateReason :: Maybe Text

The reason that the connection is in the current connection state.

connectionArn :: Maybe Text

The ARN of the connection retrieved.

secretArn :: Maybe Text

The ARN of the secret created from the authorization parameters specified for the connection.

authorizationType :: Maybe ConnectionAuthorizationType

The type of authorization specified for the connection.

name :: Maybe Text

The name of the connection retrieved.

lastModifiedTime :: Maybe POSIX

A time stamp for the time that the connection was last modified.

description :: Maybe Text

The description for the connection retrieved.

lastAuthorizedTime :: Maybe POSIX

A time stamp for the time that the connection was last authorized.

authParameters :: Maybe ConnectionAuthResponseParameters

The parameters to use for authorization for the connection.

httpStatus :: Int

The response's http status code.

newDescribeConnectionResponse

Arguments

:: Int
$sel:httpStatus:DescribeConnectionResponse'

-> DescribeConnectionResponse

Create a value of DescribeConnectionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:DescribeConnectionResponse', describeConnectionResponse_creationTime - A time stamp for the time that the connection was created.

$sel:connectionState:DescribeConnectionResponse', describeConnectionResponse_connectionState - The state of the connection retrieved.

$sel:stateReason:DescribeConnectionResponse', describeConnectionResponse_stateReason - The reason that the connection is in the current connection state.

$sel:connectionArn:DescribeConnectionResponse', describeConnectionResponse_connectionArn - The ARN of the connection retrieved.

$sel:secretArn:DescribeConnectionResponse', describeConnectionResponse_secretArn - The ARN of the secret created from the authorization parameters specified for the connection.

$sel:authorizationType:DescribeConnectionResponse', describeConnectionResponse_authorizationType - The type of authorization specified for the connection.

$sel:name:DescribeConnection', describeConnectionResponse_name - The name of the connection retrieved.

$sel:lastModifiedTime:DescribeConnectionResponse', describeConnectionResponse_lastModifiedTime - A time stamp for the time that the connection was last modified.

$sel:description:DescribeConnectionResponse', describeConnectionResponse_description - The description for the connection retrieved.

$sel:lastAuthorizedTime:DescribeConnectionResponse', describeConnectionResponse_lastAuthorizedTime - A time stamp for the time that the connection was last authorized.

$sel:authParameters:DescribeConnectionResponse', describeConnectionResponse_authParameters - The parameters to use for authorization for the connection.

$sel:httpStatus:DescribeConnectionResponse', describeConnectionResponse_httpStatus - The response's http status code.

Response Lenses

describeConnectionResponse_creationTime :: Lens' DescribeConnectionResponse (Maybe UTCTime)

A time stamp for the time that the connection was created.

describeConnectionResponse_connectionState :: Lens' DescribeConnectionResponse (Maybe ConnectionState)

The state of the connection retrieved.

describeConnectionResponse_stateReason :: Lens' DescribeConnectionResponse (Maybe Text)

The reason that the connection is in the current connection state.

describeConnectionResponse_connectionArn :: Lens' DescribeConnectionResponse (Maybe Text)

The ARN of the connection retrieved.

describeConnectionResponse_secretArn :: Lens' DescribeConnectionResponse (Maybe Text)

The ARN of the secret created from the authorization parameters specified for the connection.

describeConnectionResponse_authorizationType :: Lens' DescribeConnectionResponse (Maybe ConnectionAuthorizationType)

The type of authorization specified for the connection.

describeConnectionResponse_name :: Lens' DescribeConnectionResponse (Maybe Text)

The name of the connection retrieved.

describeConnectionResponse_lastModifiedTime :: Lens' DescribeConnectionResponse (Maybe UTCTime)

A time stamp for the time that the connection was last modified.

describeConnectionResponse_description :: Lens' DescribeConnectionResponse (Maybe Text)

The description for the connection retrieved.

describeConnectionResponse_lastAuthorizedTime :: Lens' DescribeConnectionResponse (Maybe UTCTime)

A time stamp for the time that the connection was last authorized.

describeConnectionResponse_authParameters :: Lens' DescribeConnectionResponse (Maybe ConnectionAuthResponseParameters)

The parameters to use for authorization for the connection.

describeConnectionResponse_httpStatus :: Lens' DescribeConnectionResponse Int

The response's http status code.