Network.AWS.CloudWatchEvents.DescribeEventSource

Description

This operation lists details about a partner event source that is shared with your account.

Creating a Request

data DescribeEventSource

See: newDescribeEventSource smart constructor.

Constructors

DescribeEventSource'

Fields

name :: Text

The name of the partner event source to display the details of.

newDescribeEventSource

Arguments

:: Text
$sel:name:DescribeEventSource'

-> DescribeEventSource

Create a value of DescribeEventSource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:DescribeEventSource', describeEventSource_name - The name of the partner event source to display the details of.

Request Lenses

describeEventSource_name :: Lens' DescribeEventSource Text

The name of the partner event source to display the details of.

Destructuring the Response

data DescribeEventSourceResponse

See: newDescribeEventSourceResponse smart constructor.

Constructors

DescribeEventSourceResponse'

Fields

creationTime :: Maybe POSIX

The date and time that the event source was created.

expirationTime :: Maybe POSIX

The date and time that the event source will expire if you do not create a matching event bus.

arn :: Maybe Text

The ARN of the partner event source.

name :: Maybe Text

The name of the partner event source.

state :: Maybe EventSourceState

The state of the event source. If it is ACTIVE, you have already created a matching event bus for this event source, and that event bus is active. If it is PENDING, either you haven't yet created a matching event bus, or that event bus is deactivated. If it is DELETED, you have created a matching event bus, but the event source has since been deleted.

createdBy :: Maybe Text

The name of the SaaS partner that created the event source.

httpStatus :: Int

The response's http status code.

newDescribeEventSourceResponse

Arguments

:: Int
$sel:httpStatus:DescribeEventSourceResponse'

-> DescribeEventSourceResponse

Create a value of DescribeEventSourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:DescribeEventSourceResponse', describeEventSourceResponse_creationTime - The date and time that the event source was created.

$sel:expirationTime:DescribeEventSourceResponse', describeEventSourceResponse_expirationTime - The date and time that the event source will expire if you do not create a matching event bus.

$sel:arn:DescribeEventSourceResponse', describeEventSourceResponse_arn - The ARN of the partner event source.

$sel:name:DescribeEventSource', describeEventSourceResponse_name - The name of the partner event source.

$sel:state:DescribeEventSourceResponse', describeEventSourceResponse_state - The state of the event source. If it is ACTIVE, you have already created a matching event bus for this event source, and that event bus is active. If it is PENDING, either you haven't yet created a matching event bus, or that event bus is deactivated. If it is DELETED, you have created a matching event bus, but the event source has since been deleted.

$sel:createdBy:DescribeEventSourceResponse', describeEventSourceResponse_createdBy - The name of the SaaS partner that created the event source.

$sel:httpStatus:DescribeEventSourceResponse', describeEventSourceResponse_httpStatus - The response's http status code.

Response Lenses

describeEventSourceResponse_creationTime :: Lens' DescribeEventSourceResponse (Maybe UTCTime)

The date and time that the event source was created.

describeEventSourceResponse_expirationTime :: Lens' DescribeEventSourceResponse (Maybe UTCTime)

The date and time that the event source will expire if you do not create a matching event bus.

describeEventSourceResponse_arn :: Lens' DescribeEventSourceResponse (Maybe Text)

The ARN of the partner event source.

describeEventSourceResponse_name :: Lens' DescribeEventSourceResponse (Maybe Text)

The name of the partner event source.

describeEventSourceResponse_state :: Lens' DescribeEventSourceResponse (Maybe EventSourceState)

The state of the event source. If it is ACTIVE, you have already created a matching event bus for this event source, and that event bus is active. If it is PENDING, either you haven't yet created a matching event bus, or that event bus is deactivated. If it is DELETED, you have created a matching event bus, but the event source has since been deleted.

describeEventSourceResponse_createdBy :: Lens' DescribeEventSourceResponse (Maybe Text)

The name of the SaaS partner that created the event source.

describeEventSourceResponse_httpStatus :: Lens' DescribeEventSourceResponse Int

The response's http status code.