Network.AWS.CloudWatchEvents.DescribeReplay

Description

Retrieves details about a replay. Use DescribeReplay to determine the progress of a running replay. A replay processes events to replay based on the time in the event, and replays them using 1 minute intervals. If you use StartReplay and specify an EventStartTime and an EventEndTime that covers a 20 minute time range, the events are replayed from the first minute of that 20 minute range first. Then the events from the second minute are replayed. You can use DescribeReplay to determine the progress of a replay. The value returned for EventLastReplayedTime indicates the time within the specified time range associated with the last event replayed.

Creating a Request

data DescribeReplay

See: newDescribeReplay smart constructor.

Constructors

DescribeReplay'

Fields

replayName :: Text

The name of the replay to retrieve.

newDescribeReplay

Arguments

:: Text
$sel:replayName:DescribeReplay'

-> DescribeReplay

Create a value of DescribeReplay with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:replayName:DescribeReplay', describeReplay_replayName - The name of the replay to retrieve.

Request Lenses

describeReplay_replayName :: Lens' DescribeReplay Text

The name of the replay to retrieve.

Destructuring the Response

data DescribeReplayResponse

See: newDescribeReplayResponse smart constructor.

Constructors

DescribeReplayResponse'

Fields

eventSourceArn :: Maybe Text

The ARN of the archive events were replayed from.

destination :: Maybe ReplayDestination

A ReplayDestination object that contains details about the replay.

state :: Maybe ReplayState

The current state of the replay.

eventEndTime :: Maybe POSIX

The time stamp for the last event that was replayed from the archive.

replayStartTime :: Maybe POSIX

A time stamp for the time that the replay started.

replayArn :: Maybe Text

The ARN of the replay.

replayEndTime :: Maybe POSIX

A time stamp for the time that the replay stopped.

eventLastReplayedTime :: Maybe POSIX

The time that the event was last replayed.

eventStartTime :: Maybe POSIX

The time stamp of the first event that was last replayed from the archive.

replayName :: Maybe Text

The name of the replay.

stateReason :: Maybe Text

The reason that the replay is in the current state.

description :: Maybe Text

The description of the replay.

httpStatus :: Int

The response's http status code.

newDescribeReplayResponse

Arguments

:: Int
$sel:httpStatus:DescribeReplayResponse'

-> DescribeReplayResponse

Create a value of DescribeReplayResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventSourceArn:DescribeReplayResponse', describeReplayResponse_eventSourceArn - The ARN of the archive events were replayed from.

$sel:destination:DescribeReplayResponse', describeReplayResponse_destination - A ReplayDestination object that contains details about the replay.

$sel:state:DescribeReplayResponse', describeReplayResponse_state - The current state of the replay.

$sel:eventEndTime:DescribeReplayResponse', describeReplayResponse_eventEndTime - The time stamp for the last event that was replayed from the archive.

$sel:replayStartTime:DescribeReplayResponse', describeReplayResponse_replayStartTime - A time stamp for the time that the replay started.

$sel:replayArn:DescribeReplayResponse', describeReplayResponse_replayArn - The ARN of the replay.

$sel:replayEndTime:DescribeReplayResponse', describeReplayResponse_replayEndTime - A time stamp for the time that the replay stopped.

$sel:eventLastReplayedTime:DescribeReplayResponse', describeReplayResponse_eventLastReplayedTime - The time that the event was last replayed.

$sel:eventStartTime:DescribeReplayResponse', describeReplayResponse_eventStartTime - The time stamp of the first event that was last replayed from the archive.

$sel:replayName:DescribeReplay', describeReplayResponse_replayName - The name of the replay.

$sel:stateReason:DescribeReplayResponse', describeReplayResponse_stateReason - The reason that the replay is in the current state.

$sel:description:DescribeReplayResponse', describeReplayResponse_description - The description of the replay.

$sel:httpStatus:DescribeReplayResponse', describeReplayResponse_httpStatus - The response's http status code.

Response Lenses

describeReplayResponse_eventSourceArn :: Lens' DescribeReplayResponse (Maybe Text)

The ARN of the archive events were replayed from.

describeReplayResponse_destination :: Lens' DescribeReplayResponse (Maybe ReplayDestination)

A ReplayDestination object that contains details about the replay.

describeReplayResponse_state :: Lens' DescribeReplayResponse (Maybe ReplayState)

The current state of the replay.

describeReplayResponse_eventEndTime :: Lens' DescribeReplayResponse (Maybe UTCTime)

The time stamp for the last event that was replayed from the archive.

describeReplayResponse_replayStartTime :: Lens' DescribeReplayResponse (Maybe UTCTime)

A time stamp for the time that the replay started.

describeReplayResponse_replayArn :: Lens' DescribeReplayResponse (Maybe Text)

The ARN of the replay.

describeReplayResponse_replayEndTime :: Lens' DescribeReplayResponse (Maybe UTCTime)

A time stamp for the time that the replay stopped.

describeReplayResponse_eventLastReplayedTime :: Lens' DescribeReplayResponse (Maybe UTCTime)

The time that the event was last replayed.

describeReplayResponse_eventStartTime :: Lens' DescribeReplayResponse (Maybe UTCTime)

The time stamp of the first event that was last replayed from the archive.

describeReplayResponse_replayName :: Lens' DescribeReplayResponse (Maybe Text)

The name of the replay.

describeReplayResponse_stateReason :: Lens' DescribeReplayResponse (Maybe Text)

The reason that the replay is in the current state.

describeReplayResponse_description :: Lens' DescribeReplayResponse (Maybe Text)

The description of the replay.

describeReplayResponse_httpStatus :: Lens' DescribeReplayResponse Int

The response's http status code.