Network.AWS.CloudWatchEvents.DescribeRule

Description

Describes the specified rule.

DescribeRule does not list the targets of a rule. To see the targets associated with a rule, use ListTargetsByRule.

Creating a Request

data DescribeRule

See: newDescribeRule smart constructor.

Constructors

DescribeRule'

Fields

eventBusName :: Maybe Text

The name or ARN of the event bus associated with the rule. If you omit this, the default event bus is used.

name :: Text

The name of the rule.

newDescribeRule

Arguments

:: Text
$sel:name:DescribeRule'

-> DescribeRule

Create a value of DescribeRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventBusName:DescribeRule', describeRule_eventBusName - The name or ARN of the event bus associated with the rule. If you omit this, the default event bus is used.

$sel:name:DescribeRule', describeRule_name - The name of the rule.

Request Lenses

describeRule_eventBusName :: Lens' DescribeRule (Maybe Text)

The name or ARN of the event bus associated with the rule. If you omit this, the default event bus is used.

describeRule_name :: Lens' DescribeRule Text

The name of the rule.

Destructuring the Response

data DescribeRuleResponse

See: newDescribeRuleResponse smart constructor.

Constructors

DescribeRuleResponse'

Fields

eventPattern :: Maybe Text

The event pattern. For more information, see Events and Event Patterns in the Amazon EventBridge User Guide.

roleArn :: Maybe Text

The Amazon Resource Name (ARN) of the IAM role associated with the rule.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the rule.

name :: Maybe Text

The name of the rule.

eventBusName :: Maybe Text

The name of the event bus associated with the rule.

scheduleExpression :: Maybe Text

The scheduling expression. For example, "cron(0 20 * * ? *)", "rate(5 minutes)".

state :: Maybe RuleState

Specifies whether the rule is enabled or disabled.

managedBy :: Maybe Text

If this is a managed rule, created by an Amazon Web Services service on your behalf, this field displays the principal name of the Amazon Web Services service that created the rule.

description :: Maybe Text

The description of the rule.

createdBy :: Maybe Text

The account ID of the user that created the rule. If you use PutRule to put a rule on an event bus in another account, the other account is the owner of the rule, and the rule ARN includes the account ID for that account. However, the value for CreatedBy is the account ID as the account that created the rule in the other account.

httpStatus :: Int

The response's http status code.

newDescribeRuleResponse

Arguments

:: Int
$sel:httpStatus:DescribeRuleResponse'

-> DescribeRuleResponse

Create a value of DescribeRuleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventPattern:DescribeRuleResponse', describeRuleResponse_eventPattern - The event pattern. For more information, see Events and Event Patterns in the Amazon EventBridge User Guide.

$sel:roleArn:DescribeRuleResponse', describeRuleResponse_roleArn - The Amazon Resource Name (ARN) of the IAM role associated with the rule.

$sel:arn:DescribeRuleResponse', describeRuleResponse_arn - The Amazon Resource Name (ARN) of the rule.

$sel:name:DescribeRule', describeRuleResponse_name - The name of the rule.

$sel:eventBusName:DescribeRule', describeRuleResponse_eventBusName - The name of the event bus associated with the rule.

$sel:scheduleExpression:DescribeRuleResponse', describeRuleResponse_scheduleExpression - The scheduling expression. For example, "cron(0 20 * * ? *)", "rate(5 minutes)".

$sel:state:DescribeRuleResponse', describeRuleResponse_state - Specifies whether the rule is enabled or disabled.

$sel:managedBy:DescribeRuleResponse', describeRuleResponse_managedBy - If this is a managed rule, created by an Amazon Web Services service on your behalf, this field displays the principal name of the Amazon Web Services service that created the rule.

$sel:description:DescribeRuleResponse', describeRuleResponse_description - The description of the rule.

$sel:createdBy:DescribeRuleResponse', describeRuleResponse_createdBy - The account ID of the user that created the rule. If you use PutRule to put a rule on an event bus in another account, the other account is the owner of the rule, and the rule ARN includes the account ID for that account. However, the value for CreatedBy is the account ID as the account that created the rule in the other account.

$sel:httpStatus:DescribeRuleResponse', describeRuleResponse_httpStatus - The response's http status code.

Response Lenses

describeRuleResponse_eventPattern :: Lens' DescribeRuleResponse (Maybe Text)

The event pattern. For more information, see Events and Event Patterns in the Amazon EventBridge User Guide.

describeRuleResponse_roleArn :: Lens' DescribeRuleResponse (Maybe Text)

The Amazon Resource Name (ARN) of the IAM role associated with the rule.

describeRuleResponse_arn :: Lens' DescribeRuleResponse (Maybe Text)

The Amazon Resource Name (ARN) of the rule.

describeRuleResponse_name :: Lens' DescribeRuleResponse (Maybe Text)

The name of the rule.

describeRuleResponse_eventBusName :: Lens' DescribeRuleResponse (Maybe Text)

The name of the event bus associated with the rule.

describeRuleResponse_scheduleExpression :: Lens' DescribeRuleResponse (Maybe Text)

The scheduling expression. For example, "cron(0 20 * * ? *)", "rate(5 minutes)".

describeRuleResponse_state :: Lens' DescribeRuleResponse (Maybe RuleState)

Specifies whether the rule is enabled or disabled.

describeRuleResponse_managedBy :: Lens' DescribeRuleResponse (Maybe Text)

If this is a managed rule, created by an Amazon Web Services service on your behalf, this field displays the principal name of the Amazon Web Services service that created the rule.

describeRuleResponse_description :: Lens' DescribeRuleResponse (Maybe Text)

The description of the rule.

describeRuleResponse_createdBy :: Lens' DescribeRuleResponse (Maybe Text)

The account ID of the user that created the rule. If you use PutRule to put a rule on an event bus in another account, the other account is the owner of the rule, and the rule ARN includes the account ID for that account. However, the value for CreatedBy is the account ID as the account that created the rule in the other account.

describeRuleResponse_httpStatus :: Lens' DescribeRuleResponse Int

The response's http status code.