Network.AWS.CloudWatchEvents.EnableRule

Description

Enables the specified rule. If the rule does not exist, the operation fails.

When you enable a rule, incoming events might not immediately start matching to a newly enabled rule. Allow a short period of time for changes to take effect.

Creating a Request

data EnableRule

See: newEnableRule smart constructor.

Constructors

EnableRule'

Fields

eventBusName :: Maybe Text

The name or ARN of the event bus associated with the rule. If you omit this, the default event bus is used.

name :: Text

The name of the rule.

newEnableRule

Arguments

:: Text
$sel:name:EnableRule'

-> EnableRule

Create a value of EnableRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventBusName:EnableRule', enableRule_eventBusName - The name or ARN of the event bus associated with the rule. If you omit this, the default event bus is used.

$sel:name:EnableRule', enableRule_name - The name of the rule.

Request Lenses

enableRule_eventBusName :: Lens' EnableRule (Maybe Text)

The name or ARN of the event bus associated with the rule. If you omit this, the default event bus is used.

enableRule_name :: Lens' EnableRule Text

The name of the rule.

Destructuring the Response

data EnableRuleResponse

See: newEnableRuleResponse smart constructor.

Constructors

EnableRuleResponse'

newEnableRuleResponse :: EnableRuleResponse

Create a value of EnableRuleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.