Network.AWS.CloudWatchEvents.ListApiDestinations

Description

Retrieves a list of API destination in the account in the current Region.

Creating a Request

data ListApiDestinations

See: newListApiDestinations smart constructor.

Constructors

ListApiDestinations'

Fields

nextToken :: Maybe Text

The token returned by a previous call to retrieve the next set of results.

connectionArn :: Maybe Text

The ARN of the connection specified for the API destination.

namePrefix :: Maybe Text

A name prefix to filter results returned. Only API destinations with a name that starts with the prefix are returned.

limit :: Maybe Natural

The maximum number of API destinations to include in the response.

newListApiDestinations :: ListApiDestinations

Create a value of ListApiDestinations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListApiDestinations', listApiDestinations_nextToken - The token returned by a previous call to retrieve the next set of results.

$sel:connectionArn:ListApiDestinations', listApiDestinations_connectionArn - The ARN of the connection specified for the API destination.

$sel:namePrefix:ListApiDestinations', listApiDestinations_namePrefix - A name prefix to filter results returned. Only API destinations with a name that starts with the prefix are returned.

$sel:limit:ListApiDestinations', listApiDestinations_limit - The maximum number of API destinations to include in the response.

Request Lenses

listApiDestinations_nextToken :: Lens' ListApiDestinations (Maybe Text)

The token returned by a previous call to retrieve the next set of results.

listApiDestinations_connectionArn :: Lens' ListApiDestinations (Maybe Text)

The ARN of the connection specified for the API destination.

listApiDestinations_namePrefix :: Lens' ListApiDestinations (Maybe Text)

A name prefix to filter results returned. Only API destinations with a name that starts with the prefix are returned.

listApiDestinations_limit :: Lens' ListApiDestinations (Maybe Natural)

The maximum number of API destinations to include in the response.

Destructuring the Response

data ListApiDestinationsResponse

See: newListApiDestinationsResponse smart constructor.

Constructors

ListApiDestinationsResponse'

Fields

nextToken :: Maybe Text

A token you can use in a subsequent request to retrieve the next set of results.

apiDestinations :: Maybe [ApiDestination]

An array of ApiDestination objects that include information about an API destination.

httpStatus :: Int

The response's http status code.

newListApiDestinationsResponse

Arguments

:: Int
$sel:httpStatus:ListApiDestinationsResponse'

-> ListApiDestinationsResponse

Create a value of ListApiDestinationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListApiDestinations', listApiDestinationsResponse_nextToken - A token you can use in a subsequent request to retrieve the next set of results.

$sel:apiDestinations:ListApiDestinationsResponse', listApiDestinationsResponse_apiDestinations - An array of ApiDestination objects that include information about an API destination.

$sel:httpStatus:ListApiDestinationsResponse', listApiDestinationsResponse_httpStatus - The response's http status code.

Response Lenses

listApiDestinationsResponse_nextToken :: Lens' ListApiDestinationsResponse (Maybe Text)

A token you can use in a subsequent request to retrieve the next set of results.

listApiDestinationsResponse_apiDestinations :: Lens' ListApiDestinationsResponse (Maybe [ApiDestination])

An array of ApiDestination objects that include information about an API destination.

listApiDestinationsResponse_httpStatus :: Lens' ListApiDestinationsResponse Int

The response's http status code.