Network.AWS.CloudWatchEvents.ListPartnerEventSourceAccounts

Description

An SaaS partner can use this operation to display the Amazon Web Services account ID that a particular partner event source name is associated with. This operation is not used by Amazon Web Services customers.

Creating a Request

data ListPartnerEventSourceAccounts

See: newListPartnerEventSourceAccounts smart constructor.

Constructors

ListPartnerEventSourceAccounts'

Fields

nextToken :: Maybe Text

The token returned by a previous call to this operation. Specifying this retrieves the next set of results.

limit :: Maybe Natural

Specifying this limits the number of results returned by this operation. The operation also returns a NextToken which you can use in a subsequent operation to retrieve the next set of results.

eventSourceName :: Text

The name of the partner event source to display account information about.

newListPartnerEventSourceAccounts

Arguments

:: Text
$sel:eventSourceName:ListPartnerEventSourceAccounts'

-> ListPartnerEventSourceAccounts

Create a value of ListPartnerEventSourceAccounts with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListPartnerEventSourceAccounts', listPartnerEventSourceAccounts_nextToken - The token returned by a previous call to this operation. Specifying this retrieves the next set of results.

$sel:limit:ListPartnerEventSourceAccounts', listPartnerEventSourceAccounts_limit - Specifying this limits the number of results returned by this operation. The operation also returns a NextToken which you can use in a subsequent operation to retrieve the next set of results.

$sel:eventSourceName:ListPartnerEventSourceAccounts', listPartnerEventSourceAccounts_eventSourceName - The name of the partner event source to display account information about.

Request Lenses

listPartnerEventSourceAccounts_nextToken :: Lens' ListPartnerEventSourceAccounts (Maybe Text)

The token returned by a previous call to this operation. Specifying this retrieves the next set of results.

listPartnerEventSourceAccounts_limit :: Lens' ListPartnerEventSourceAccounts (Maybe Natural)

Specifying this limits the number of results returned by this operation. The operation also returns a NextToken which you can use in a subsequent operation to retrieve the next set of results.

listPartnerEventSourceAccounts_eventSourceName :: Lens' ListPartnerEventSourceAccounts Text

The name of the partner event source to display account information about.

Destructuring the Response

data ListPartnerEventSourceAccountsResponse

See: newListPartnerEventSourceAccountsResponse smart constructor.

Constructors

ListPartnerEventSourceAccountsResponse'

Fields

nextToken :: Maybe Text

A token you can use in a subsequent operation to retrieve the next set of results.

partnerEventSourceAccounts :: Maybe [PartnerEventSourceAccount]

The list of partner event sources returned by the operation.

httpStatus :: Int

The response's http status code.

newListPartnerEventSourceAccountsResponse

Arguments

:: Int
$sel:httpStatus:ListPartnerEventSourceAccountsResponse'

-> ListPartnerEventSourceAccountsResponse

Create a value of ListPartnerEventSourceAccountsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListPartnerEventSourceAccounts', listPartnerEventSourceAccountsResponse_nextToken - A token you can use in a subsequent operation to retrieve the next set of results.

$sel:partnerEventSourceAccounts:ListPartnerEventSourceAccountsResponse', listPartnerEventSourceAccountsResponse_partnerEventSourceAccounts - The list of partner event sources returned by the operation.

$sel:httpStatus:ListPartnerEventSourceAccountsResponse', listPartnerEventSourceAccountsResponse_httpStatus - The response's http status code.

Response Lenses

listPartnerEventSourceAccountsResponse_nextToken :: Lens' ListPartnerEventSourceAccountsResponse (Maybe Text)

A token you can use in a subsequent operation to retrieve the next set of results.

listPartnerEventSourceAccountsResponse_partnerEventSourceAccounts :: Lens' ListPartnerEventSourceAccountsResponse (Maybe [PartnerEventSourceAccount])

The list of partner event sources returned by the operation.

listPartnerEventSourceAccountsResponse_httpStatus :: Lens' ListPartnerEventSourceAccountsResponse Int

The response's http status code.