Network.AWS.CloudWatchEvents.ListTargetsByRule

Description

Lists the targets assigned to the specified rule.

This operation returns paginated results.

Creating a Request

data ListTargetsByRule

See: newListTargetsByRule smart constructor.

Constructors

ListTargetsByRule'

Fields

nextToken :: Maybe Text

The token returned by a previous call to retrieve the next set of results.

eventBusName :: Maybe Text

The name or ARN of the event bus associated with the rule. If you omit this, the default event bus is used.

limit :: Maybe Natural

The maximum number of results to return.

rule :: Text

The name of the rule.

newListTargetsByRule

Arguments

:: Text
$sel:rule:ListTargetsByRule'

-> ListTargetsByRule

Create a value of ListTargetsByRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListTargetsByRule', listTargetsByRule_nextToken - The token returned by a previous call to retrieve the next set of results.

$sel:eventBusName:ListTargetsByRule', listTargetsByRule_eventBusName - The name or ARN of the event bus associated with the rule. If you omit this, the default event bus is used.

$sel:limit:ListTargetsByRule', listTargetsByRule_limit - The maximum number of results to return.

$sel:rule:ListTargetsByRule', listTargetsByRule_rule - The name of the rule.

Request Lenses

listTargetsByRule_nextToken :: Lens' ListTargetsByRule (Maybe Text)

The token returned by a previous call to retrieve the next set of results.

listTargetsByRule_eventBusName :: Lens' ListTargetsByRule (Maybe Text)

The name or ARN of the event bus associated with the rule. If you omit this, the default event bus is used.

listTargetsByRule_limit :: Lens' ListTargetsByRule (Maybe Natural)

The maximum number of results to return.

listTargetsByRule_rule :: Lens' ListTargetsByRule Text

The name of the rule.

Destructuring the Response

data ListTargetsByRuleResponse

See: newListTargetsByRuleResponse smart constructor.

Constructors

ListTargetsByRuleResponse'

Fields

nextToken :: Maybe Text

Indicates whether there are additional results to retrieve. If there are no more results, the value is null.

targets :: Maybe (NonEmpty Target)

The targets assigned to the rule.

httpStatus :: Int

The response's http status code.

newListTargetsByRuleResponse

Arguments

:: Int
$sel:httpStatus:ListTargetsByRuleResponse'

-> ListTargetsByRuleResponse

Create a value of ListTargetsByRuleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListTargetsByRule', listTargetsByRuleResponse_nextToken - Indicates whether there are additional results to retrieve. If there are no more results, the value is null.

$sel:targets:ListTargetsByRuleResponse', listTargetsByRuleResponse_targets - The targets assigned to the rule.

$sel:httpStatus:ListTargetsByRuleResponse', listTargetsByRuleResponse_httpStatus - The response's http status code.

Response Lenses

listTargetsByRuleResponse_nextToken :: Lens' ListTargetsByRuleResponse (Maybe Text)

Indicates whether there are additional results to retrieve. If there are no more results, the value is null.

listTargetsByRuleResponse_targets :: Lens' ListTargetsByRuleResponse (Maybe (NonEmpty Target))

The targets assigned to the rule.

listTargetsByRuleResponse_httpStatus :: Lens' ListTargetsByRuleResponse Int

The response's http status code.