Network.AWS.CloudWatchEvents.RemoveTargets

Description

Removes the specified targets from the specified rule. When the rule is triggered, those targets are no longer be invoked.

When you remove a target, when the associated rule triggers, removed targets might continue to be invoked. Allow a short period of time for changes to take effect.

This action can partially fail if too many requests are made at the same time. If that happens, FailedEntryCount is non-zero in the response and each entry in FailedEntries provides the ID of the failed target and the error code.

Creating a Request

data RemoveTargets

See: newRemoveTargets smart constructor.

Constructors

RemoveTargets'

Fields

force :: Maybe Bool

If this is a managed rule, created by an Amazon Web Services service on your behalf, you must specify Force as True to remove targets. This parameter is ignored for rules that are not managed rules. You can check whether a rule is a managed rule by using DescribeRule or ListRules and checking the ManagedBy field of the response.

eventBusName :: Maybe Text

The name or ARN of the event bus associated with the rule. If you omit this, the default event bus is used.

rule :: Text

The name of the rule.

ids :: NonEmpty Text

The IDs of the targets to remove from the rule.

newRemoveTargets

Arguments

:: Text
$sel:rule:RemoveTargets'

-> NonEmpty Text
$sel:ids:RemoveTargets'

-> RemoveTargets

Create a value of RemoveTargets with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:force:RemoveTargets', removeTargets_force - If this is a managed rule, created by an Amazon Web Services service on your behalf, you must specify Force as True to remove targets. This parameter is ignored for rules that are not managed rules. You can check whether a rule is a managed rule by using DescribeRule or ListRules and checking the ManagedBy field of the response.

$sel:eventBusName:RemoveTargets', removeTargets_eventBusName - The name or ARN of the event bus associated with the rule. If you omit this, the default event bus is used.

$sel:rule:RemoveTargets', removeTargets_rule - The name of the rule.

$sel:ids:RemoveTargets', removeTargets_ids - The IDs of the targets to remove from the rule.

Request Lenses

removeTargets_force :: Lens' RemoveTargets (Maybe Bool)

If this is a managed rule, created by an Amazon Web Services service on your behalf, you must specify Force as True to remove targets. This parameter is ignored for rules that are not managed rules. You can check whether a rule is a managed rule by using DescribeRule or ListRules and checking the ManagedBy field of the response.

removeTargets_eventBusName :: Lens' RemoveTargets (Maybe Text)

The name or ARN of the event bus associated with the rule. If you omit this, the default event bus is used.

removeTargets_rule :: Lens' RemoveTargets Text

The name of the rule.

removeTargets_ids :: Lens' RemoveTargets (NonEmpty Text)

The IDs of the targets to remove from the rule.

Destructuring the Response

data RemoveTargetsResponse

See: newRemoveTargetsResponse smart constructor.

Constructors

RemoveTargetsResponse'

Fields

failedEntryCount :: Maybe Int

The number of failed entries.

failedEntries :: Maybe [RemoveTargetsResultEntry]

The failed target entries.

httpStatus :: Int

The response's http status code.

newRemoveTargetsResponse

Arguments

:: Int
$sel:httpStatus:RemoveTargetsResponse'

-> RemoveTargetsResponse

Create a value of RemoveTargetsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failedEntryCount:RemoveTargetsResponse', removeTargetsResponse_failedEntryCount - The number of failed entries.

$sel:failedEntries:RemoveTargetsResponse', removeTargetsResponse_failedEntries - The failed target entries.

$sel:httpStatus:RemoveTargetsResponse', removeTargetsResponse_httpStatus - The response's http status code.

Response Lenses

removeTargetsResponse_failedEntryCount :: Lens' RemoveTargetsResponse (Maybe Int)

The number of failed entries.

removeTargetsResponse_failedEntries :: Lens' RemoveTargetsResponse (Maybe [RemoveTargetsResultEntry])

The failed target entries.

removeTargetsResponse_httpStatus :: Lens' RemoveTargetsResponse Int

The response's http status code.