Network.AWS.CloudWatchEvents.StartReplay

Description

Starts the specified replay. Events are not necessarily replayed in the exact same order that they were added to the archive. A replay processes events to replay based on the time in the event, and replays them using 1 minute intervals. If you specify an EventStartTime and an EventEndTime that covers a 20 minute time range, the events are replayed from the first minute of that 20 minute range first. Then the events from the second minute are replayed. You can use DescribeReplay to determine the progress of a replay. The value returned for EventLastReplayedTime indicates the time within the specified time range associated with the last event replayed.

Creating a Request

data StartReplay

See: newStartReplay smart constructor.

Constructors

StartReplay'

Fields

description :: Maybe Text

A description for the replay to start.

replayName :: Text

The name of the replay to start.

eventSourceArn :: Text

The ARN of the archive to replay events from.

eventStartTime :: POSIX

A time stamp for the time to start replaying events. Only events that occurred between the EventStartTime and EventEndTime are replayed.

eventEndTime :: POSIX

A time stamp for the time to stop replaying events. Only events that occurred between the EventStartTime and EventEndTime are replayed.

destination :: ReplayDestination

A ReplayDestination object that includes details about the destination for the replay.

newStartReplay

Arguments

:: Text
$sel:replayName:StartReplay'

-> Text
$sel:eventSourceArn:StartReplay'

-> UTCTime
$sel:eventStartTime:StartReplay'

-> UTCTime
$sel:eventEndTime:StartReplay'

-> ReplayDestination
$sel:destination:StartReplay'

-> StartReplay

Create a value of StartReplay with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:StartReplay', startReplay_description - A description for the replay to start.

$sel:replayName:StartReplay', startReplay_replayName - The name of the replay to start.

$sel:eventSourceArn:StartReplay', startReplay_eventSourceArn - The ARN of the archive to replay events from.

$sel:eventStartTime:StartReplay', startReplay_eventStartTime - A time stamp for the time to start replaying events. Only events that occurred between the EventStartTime and EventEndTime are replayed.

$sel:eventEndTime:StartReplay', startReplay_eventEndTime - A time stamp for the time to stop replaying events. Only events that occurred between the EventStartTime and EventEndTime are replayed.

$sel:destination:StartReplay', startReplay_destination - A ReplayDestination object that includes details about the destination for the replay.

Request Lenses

startReplay_description :: Lens' StartReplay (Maybe Text)

A description for the replay to start.

startReplay_replayName :: Lens' StartReplay Text

The name of the replay to start.

startReplay_eventSourceArn :: Lens' StartReplay Text

The ARN of the archive to replay events from.

startReplay_eventStartTime :: Lens' StartReplay UTCTime

A time stamp for the time to start replaying events. Only events that occurred between the EventStartTime and EventEndTime are replayed.

startReplay_eventEndTime :: Lens' StartReplay UTCTime

A time stamp for the time to stop replaying events. Only events that occurred between the EventStartTime and EventEndTime are replayed.

startReplay_destination :: Lens' StartReplay ReplayDestination

A ReplayDestination object that includes details about the destination for the replay.

Destructuring the Response

data StartReplayResponse

See: newStartReplayResponse smart constructor.

Constructors

StartReplayResponse'

Fields

replayArn :: Maybe Text

The ARN of the replay.

replayStartTime :: Maybe POSIX

The time at which the replay started.

stateReason :: Maybe Text

The reason that the replay is in the state.

state :: Maybe ReplayState

The state of the replay.

httpStatus :: Int

The response's http status code.

newStartReplayResponse

Arguments

:: Int
$sel:httpStatus:StartReplayResponse'

-> StartReplayResponse

Create a value of StartReplayResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:replayArn:StartReplayResponse', startReplayResponse_replayArn - The ARN of the replay.

$sel:replayStartTime:StartReplayResponse', startReplayResponse_replayStartTime - The time at which the replay started.

$sel:stateReason:StartReplayResponse', startReplayResponse_stateReason - The reason that the replay is in the state.

$sel:state:StartReplayResponse', startReplayResponse_state - The state of the replay.

$sel:httpStatus:StartReplayResponse', startReplayResponse_httpStatus - The response's http status code.

Response Lenses

startReplayResponse_replayArn :: Lens' StartReplayResponse (Maybe Text)

The ARN of the replay.

startReplayResponse_replayStartTime :: Lens' StartReplayResponse (Maybe UTCTime)

The time at which the replay started.

startReplayResponse_stateReason :: Lens' StartReplayResponse (Maybe Text)

The reason that the replay is in the state.

startReplayResponse_state :: Lens' StartReplayResponse (Maybe ReplayState)

The state of the replay.

startReplayResponse_httpStatus :: Lens' StartReplayResponse Int

The response's http status code.