Network.AWS.CloudWatchEvents.TestEventPattern

Description

Tests whether the specified event pattern matches the provided event.

Most services in Amazon Web Services treat : or / as the same character in Amazon Resource Names (ARNs). However, EventBridge uses an exact match in event patterns and rules. Be sure to use the correct ARN characters when creating event patterns so that they match the ARN syntax in the event you want to match.

Creating a Request

data TestEventPattern

See: newTestEventPattern smart constructor.

Constructors

TestEventPattern'

Fields

eventPattern :: Text

The event pattern. For more information, see Events and Event Patterns in the Amazon EventBridge User Guide.

event :: Text

The event, in JSON format, to test against the event pattern. The JSON must follow the format specified in Amazon Web Services Events, and the following fields are mandatory:

id

account

source

time

region

resources

detail-type

newTestEventPattern

Arguments

:: Text
$sel:eventPattern:TestEventPattern'

-> Text
$sel:event:TestEventPattern'

-> TestEventPattern

Create a value of TestEventPattern with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventPattern:TestEventPattern', testEventPattern_eventPattern - The event pattern. For more information, see Events and Event Patterns in the Amazon EventBridge User Guide.

$sel:event:TestEventPattern', testEventPattern_event - The event, in JSON format, to test against the event pattern. The JSON must follow the format specified in Amazon Web Services Events, and the following fields are mandatory:

id

account

source

time

region

resources

detail-type

Request Lenses

testEventPattern_eventPattern :: Lens' TestEventPattern Text

The event pattern. For more information, see Events and Event Patterns in the Amazon EventBridge User Guide.

testEventPattern_event :: Lens' TestEventPattern Text

The event, in JSON format, to test against the event pattern. The JSON must follow the format specified in Amazon Web Services Events, and the following fields are mandatory:

id

account

source

time

region

resources

detail-type

Destructuring the Response

data TestEventPatternResponse

See: newTestEventPatternResponse smart constructor.

Constructors

TestEventPatternResponse'

Fields

result :: Maybe Bool

Indicates whether the event matches the event pattern.

httpStatus :: Int

The response's http status code.

newTestEventPatternResponse

Arguments

:: Int
$sel:httpStatus:TestEventPatternResponse'

-> TestEventPatternResponse

Create a value of TestEventPatternResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:result:TestEventPatternResponse', testEventPatternResponse_result - Indicates whether the event matches the event pattern.

$sel:httpStatus:TestEventPatternResponse', testEventPatternResponse_httpStatus - The response's http status code.

Response Lenses

testEventPatternResponse_result :: Lens' TestEventPatternResponse (Maybe Bool)

Indicates whether the event matches the event pattern.

testEventPatternResponse_httpStatus :: Lens' TestEventPatternResponse Int

The response's http status code.