Network.AWS.CloudWatchEvents.Types.ApiDestination

Description

Documentation

data ApiDestination

Contains details about an API destination.

See: newApiDestination smart constructor.

Constructors

ApiDestination'

Fields

httpMethod :: Maybe ApiDestinationHttpMethod

The method to use to connect to the HTTP endpoint.

creationTime :: Maybe POSIX

A time stamp for the time that the API destination was created.

apiDestinationArn :: Maybe Text

The ARN of the API destination.

invocationEndpoint :: Maybe Text

The URL to the endpoint for the API destination.

connectionArn :: Maybe Text

The ARN of the connection specified for the API destination.

apiDestinationState :: Maybe ApiDestinationState

The state of the API destination.

name :: Maybe Text

The name of the API destination.

lastModifiedTime :: Maybe POSIX

A time stamp for the time that the API destination was last modified.

invocationRateLimitPerSecond :: Maybe Natural

The maximum number of invocations per second to send to the HTTP endpoint.

newApiDestination :: ApiDestination

Create a value of ApiDestination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpMethod:ApiDestination', apiDestination_httpMethod - The method to use to connect to the HTTP endpoint.

$sel:creationTime:ApiDestination', apiDestination_creationTime - A time stamp for the time that the API destination was created.

$sel:apiDestinationArn:ApiDestination', apiDestination_apiDestinationArn - The ARN of the API destination.

$sel:invocationEndpoint:ApiDestination', apiDestination_invocationEndpoint - The URL to the endpoint for the API destination.

$sel:connectionArn:ApiDestination', apiDestination_connectionArn - The ARN of the connection specified for the API destination.

$sel:apiDestinationState:ApiDestination', apiDestination_apiDestinationState - The state of the API destination.

$sel:name:ApiDestination', apiDestination_name - The name of the API destination.

$sel:lastModifiedTime:ApiDestination', apiDestination_lastModifiedTime - A time stamp for the time that the API destination was last modified.

$sel:invocationRateLimitPerSecond:ApiDestination', apiDestination_invocationRateLimitPerSecond - The maximum number of invocations per second to send to the HTTP endpoint.

apiDestination_httpMethod :: Lens' ApiDestination (Maybe ApiDestinationHttpMethod)

The method to use to connect to the HTTP endpoint.

apiDestination_creationTime :: Lens' ApiDestination (Maybe UTCTime)

A time stamp for the time that the API destination was created.

apiDestination_apiDestinationArn :: Lens' ApiDestination (Maybe Text)

The ARN of the API destination.

apiDestination_invocationEndpoint :: Lens' ApiDestination (Maybe Text)

The URL to the endpoint for the API destination.

apiDestination_connectionArn :: Lens' ApiDestination (Maybe Text)

The ARN of the connection specified for the API destination.

apiDestination_apiDestinationState :: Lens' ApiDestination (Maybe ApiDestinationState)

The state of the API destination.

apiDestination_name :: Lens' ApiDestination (Maybe Text)

The name of the API destination.

apiDestination_lastModifiedTime :: Lens' ApiDestination (Maybe UTCTime)

A time stamp for the time that the API destination was last modified.

apiDestination_invocationRateLimitPerSecond :: Lens' ApiDestination (Maybe Natural)

The maximum number of invocations per second to send to the HTTP endpoint.