Network.AWS.CloudWatchEvents.Types.Archive

Description

Documentation

data Archive

An Archive object that contains details about an archive.

See: newArchive smart constructor.

Constructors

Archive'

Fields

eventCount :: Maybe Integer

The number of events in the archive.

creationTime :: Maybe POSIX

The time stamp for the time that the archive was created.

eventSourceArn :: Maybe Text

The ARN of the event bus associated with the archive. Only events from this event bus are sent to the archive.

stateReason :: Maybe Text

A description for the reason that the archive is in the current state.

archiveName :: Maybe Text

The name of the archive.

state :: Maybe ArchiveState

The current state of the archive.

sizeBytes :: Maybe Integer

The size of the archive, in bytes.

retentionDays :: Maybe Natural

The number of days to retain events in the archive before they are deleted.

newArchive :: Archive

Create a value of Archive with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventCount:Archive', archive_eventCount - The number of events in the archive.

$sel:creationTime:Archive', archive_creationTime - The time stamp for the time that the archive was created.

$sel:eventSourceArn:Archive', archive_eventSourceArn - The ARN of the event bus associated with the archive. Only events from this event bus are sent to the archive.

$sel:stateReason:Archive', archive_stateReason - A description for the reason that the archive is in the current state.

$sel:archiveName:Archive', archive_archiveName - The name of the archive.

$sel:state:Archive', archive_state - The current state of the archive.

$sel:sizeBytes:Archive', archive_sizeBytes - The size of the archive, in bytes.

$sel:retentionDays:Archive', archive_retentionDays - The number of days to retain events in the archive before they are deleted.

archive_eventCount :: Lens' Archive (Maybe Integer)

The number of events in the archive.

archive_creationTime :: Lens' Archive (Maybe UTCTime)

The time stamp for the time that the archive was created.

archive_eventSourceArn :: Lens' Archive (Maybe Text)

The ARN of the event bus associated with the archive. Only events from this event bus are sent to the archive.

archive_stateReason :: Lens' Archive (Maybe Text)

A description for the reason that the archive is in the current state.

archive_archiveName :: Lens' Archive (Maybe Text)

The name of the archive.

archive_state :: Lens' Archive (Maybe ArchiveState)

The current state of the archive.

archive_sizeBytes :: Lens' Archive (Maybe Integer)

The size of the archive, in bytes.

archive_retentionDays :: Lens' Archive (Maybe Natural)

The number of days to retain events in the archive before they are deleted.