Network.AWS.CloudWatchEvents.Types.Condition

Description

Documentation

data Condition

A JSON string which you can use to limit the event bus permissions you are granting to only accounts that fulfill the condition. Currently, the only supported condition is membership in a certain Amazon Web Services organization. The string must contain Type, Key, and Value fields. The Value field specifies the ID of the Amazon Web Services organization. Following is an example value for Condition:

'{"Type" : "StringEquals", "Key": "aws:PrincipalOrgID", "Value": "o-1234567890"}'

See: newCondition smart constructor.

Constructors

Condition'

Fields

type' :: Text

Specifies the type of condition. Currently the only supported value is StringEquals.

key :: Text

Specifies the key for the condition. Currently the only supported key is aws:PrincipalOrgID.

value :: Text

Specifies the value for the key. Currently, this must be the ID of the organization.

newCondition

Arguments

:: Text
$sel:type':Condition'

-> Text
$sel:key:Condition'

-> Text
$sel:value:Condition'

-> Condition

Create a value of Condition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:type':Condition', condition_type - Specifies the type of condition. Currently the only supported value is StringEquals.

$sel:key:Condition', condition_key - Specifies the key for the condition. Currently the only supported key is aws:PrincipalOrgID.

$sel:value:Condition', condition_value - Specifies the value for the key. Currently, this must be the ID of the organization.

condition_type :: Lens' Condition Text

Specifies the type of condition. Currently the only supported value is StringEquals.

condition_key :: Lens' Condition Text

Specifies the key for the condition. Currently the only supported key is aws:PrincipalOrgID.

condition_value :: Lens' Condition Text

Specifies the value for the key. Currently, this must be the ID of the organization.