Network.AWS.CloudWatchEvents.Types.Connection

Description

Documentation

data Connection

Contains information about a connection.

See: newConnection smart constructor.

Constructors

Connection'

Fields

creationTime :: Maybe POSIX

A time stamp for the time that the connection was created.

lastModifiedTime :: Maybe POSIX

A time stamp for the time that the connection was last modified.

name :: Maybe Text

The name of the connection.

lastAuthorizedTime :: Maybe POSIX

A time stamp for the time that the connection was last authorized.

authorizationType :: Maybe ConnectionAuthorizationType

The authorization type specified for the connection.

connectionArn :: Maybe Text

The ARN of the connection.

stateReason :: Maybe Text

The reason that the connection is in the connection state.

connectionState :: Maybe ConnectionState

The state of the connection.

newConnection :: Connection

Create a value of Connection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:Connection', connection_creationTime - A time stamp for the time that the connection was created.

$sel:lastModifiedTime:Connection', connection_lastModifiedTime - A time stamp for the time that the connection was last modified.

$sel:name:Connection', connection_name - The name of the connection.

$sel:lastAuthorizedTime:Connection', connection_lastAuthorizedTime - A time stamp for the time that the connection was last authorized.

$sel:authorizationType:Connection', connection_authorizationType - The authorization type specified for the connection.

$sel:connectionArn:Connection', connection_connectionArn - The ARN of the connection.

$sel:stateReason:Connection', connection_stateReason - The reason that the connection is in the connection state.

$sel:connectionState:Connection', connection_connectionState - The state of the connection.

connection_creationTime :: Lens' Connection (Maybe UTCTime)

A time stamp for the time that the connection was created.

connection_lastModifiedTime :: Lens' Connection (Maybe UTCTime)

A time stamp for the time that the connection was last modified.

connection_name :: Lens' Connection (Maybe Text)

The name of the connection.

connection_lastAuthorizedTime :: Lens' Connection (Maybe UTCTime)

A time stamp for the time that the connection was last authorized.

connection_authorizationType :: Lens' Connection (Maybe ConnectionAuthorizationType)

The authorization type specified for the connection.

connection_connectionArn :: Lens' Connection (Maybe Text)

The ARN of the connection.

connection_stateReason :: Lens' Connection (Maybe Text)

The reason that the connection is in the connection state.

connection_connectionState :: Lens' Connection (Maybe ConnectionState)

The state of the connection.