Network.AWS.CloudWatchEvents.Types.ConnectionAuthResponseParameters

Description

Documentation

data ConnectionAuthResponseParameters

Contains the authorization parameters to use for the connection.

See: newConnectionAuthResponseParameters smart constructor.

Constructors

ConnectionAuthResponseParameters'

Fields

basicAuthParameters :: Maybe ConnectionBasicAuthResponseParameters

The authorization parameters for Basic authorization.

apiKeyAuthParameters :: Maybe ConnectionApiKeyAuthResponseParameters

The API Key parameters to use for authorization.

oAuthParameters :: Maybe ConnectionOAuthResponseParameters

The OAuth parameters to use for authorization.

invocationHttpParameters :: Maybe ConnectionHttpParameters

Additional parameters for the connection that are passed through with every invocation to the HTTP endpoint.

newConnectionAuthResponseParameters :: ConnectionAuthResponseParameters

Create a value of ConnectionAuthResponseParameters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:basicAuthParameters:ConnectionAuthResponseParameters', connectionAuthResponseParameters_basicAuthParameters - The authorization parameters for Basic authorization.

$sel:apiKeyAuthParameters:ConnectionAuthResponseParameters', connectionAuthResponseParameters_apiKeyAuthParameters - The API Key parameters to use for authorization.

$sel:oAuthParameters:ConnectionAuthResponseParameters', connectionAuthResponseParameters_oAuthParameters - The OAuth parameters to use for authorization.

$sel:invocationHttpParameters:ConnectionAuthResponseParameters', connectionAuthResponseParameters_invocationHttpParameters - Additional parameters for the connection that are passed through with every invocation to the HTTP endpoint.

connectionAuthResponseParameters_basicAuthParameters :: Lens' ConnectionAuthResponseParameters (Maybe ConnectionBasicAuthResponseParameters)

The authorization parameters for Basic authorization.

connectionAuthResponseParameters_apiKeyAuthParameters :: Lens' ConnectionAuthResponseParameters (Maybe ConnectionApiKeyAuthResponseParameters)

The API Key parameters to use for authorization.

connectionAuthResponseParameters_oAuthParameters :: Lens' ConnectionAuthResponseParameters (Maybe ConnectionOAuthResponseParameters)

The OAuth parameters to use for authorization.

connectionAuthResponseParameters_invocationHttpParameters :: Lens' ConnectionAuthResponseParameters (Maybe ConnectionHttpParameters)

Additional parameters for the connection that are passed through with every invocation to the HTTP endpoint.