Network.AWS.CloudWatchEvents.Types.EcsParameters

Description

Documentation

data EcsParameters

The custom parameters to be used when the target is an Amazon ECS task.

See: newEcsParameters smart constructor.

Constructors

EcsParameters'

Fields

group' :: Maybe Text

Specifies an ECS task group for the task. The maximum length is 255 characters.

propagateTags :: Maybe PropagateTags

Specifies whether to propagate the tags from the task definition to the task. If no value is specified, the tags are not propagated. Tags can only be propagated to the task during task creation. To add tags to a task after task creation, use the TagResource API action.

platformVersion :: Maybe Text

Specifies the platform version for the task. Specify only the numeric portion of the platform version, such as 1.1.0.

This structure is used only if LaunchType is FARGATE. For more information about valid platform versions, see Fargate Platform Versions in the Amazon Elastic Container Service Developer Guide.

enableECSManagedTags :: Maybe Bool

Specifies whether to enable Amazon ECS managed tags for the task. For more information, see Tagging Your Amazon ECS Resources in the Amazon Elastic Container Service Developer Guide.

referenceId :: Maybe Text

The reference ID to use for the task.

placementConstraints :: Maybe [PlacementConstraint]

An array of placement constraint objects to use for the task. You can specify up to 10 constraints per task (including constraints in the task definition and those specified at runtime).

placementStrategy :: Maybe [PlacementStrategy]

The placement strategy objects to use for the task. You can specify a maximum of five strategy rules per task.

launchType :: Maybe LaunchType

Specifies the launch type on which your task is running. The launch type that you specify here must match one of the launch type (compatibilities) of the target task. The FARGATE value is supported only in the Regions where Fargate witt Amazon ECS is supported. For more information, see Fargate on Amazon ECS in the Amazon Elastic Container Service Developer Guide.

capacityProviderStrategy :: Maybe [CapacityProviderStrategyItem]

The capacity provider strategy to use for the task.

If a capacityProviderStrategy is specified, the launchType parameter must be omitted. If no capacityProviderStrategy or launchType is specified, the defaultCapacityProviderStrategy for the cluster is used.

taskCount :: Maybe Natural

The number of tasks to create based on TaskDefinition. The default is 1.

networkConfiguration :: Maybe NetworkConfiguration

Use this structure if the Amazon ECS task uses the awsvpc network mode. This structure specifies the VPC subnets and security groups associated with the task, and whether a public IP address is to be used. This structure is required if LaunchType is FARGATE because the awsvpc mode is required for Fargate tasks.

If you specify NetworkConfiguration when the target ECS task does not use the awsvpc network mode, the task fails.

tags :: Maybe [Tag]

The metadata that you apply to the task to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define. To learn more, see RunTask in the Amazon ECS API Reference.

enableExecuteCommand :: Maybe Bool

Whether or not to enable the execute command functionality for the containers in this task. If true, this enables execute command functionality on all containers in the task.

taskDefinitionArn :: Text

The ARN of the task definition to use if the event target is an Amazon ECS task.

newEcsParameters

Arguments

:: Text
$sel:taskDefinitionArn:EcsParameters'

-> EcsParameters

Create a value of EcsParameters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:group':EcsParameters', ecsParameters_group - Specifies an ECS task group for the task. The maximum length is 255 characters.

$sel:propagateTags:EcsParameters', ecsParameters_propagateTags - Specifies whether to propagate the tags from the task definition to the task. If no value is specified, the tags are not propagated. Tags can only be propagated to the task during task creation. To add tags to a task after task creation, use the TagResource API action.

$sel:platformVersion:EcsParameters', ecsParameters_platformVersion - Specifies the platform version for the task. Specify only the numeric portion of the platform version, such as 1.1.0.

This structure is used only if LaunchType is FARGATE. For more information about valid platform versions, see Fargate Platform Versions in the Amazon Elastic Container Service Developer Guide.

$sel:enableECSManagedTags:EcsParameters', ecsParameters_enableECSManagedTags - Specifies whether to enable Amazon ECS managed tags for the task. For more information, see Tagging Your Amazon ECS Resources in the Amazon Elastic Container Service Developer Guide.

$sel:referenceId:EcsParameters', ecsParameters_referenceId - The reference ID to use for the task.

$sel:placementConstraints:EcsParameters', ecsParameters_placementConstraints - An array of placement constraint objects to use for the task. You can specify up to 10 constraints per task (including constraints in the task definition and those specified at runtime).

$sel:placementStrategy:EcsParameters', ecsParameters_placementStrategy - The placement strategy objects to use for the task. You can specify a maximum of five strategy rules per task.

$sel:launchType:EcsParameters', ecsParameters_launchType - Specifies the launch type on which your task is running. The launch type that you specify here must match one of the launch type (compatibilities) of the target task. The FARGATE value is supported only in the Regions where Fargate witt Amazon ECS is supported. For more information, see Fargate on Amazon ECS in the Amazon Elastic Container Service Developer Guide.

$sel:capacityProviderStrategy:EcsParameters', ecsParameters_capacityProviderStrategy - The capacity provider strategy to use for the task.

If a capacityProviderStrategy is specified, the launchType parameter must be omitted. If no capacityProviderStrategy or launchType is specified, the defaultCapacityProviderStrategy for the cluster is used.

$sel:taskCount:EcsParameters', ecsParameters_taskCount - The number of tasks to create based on TaskDefinition. The default is 1.

$sel:networkConfiguration:EcsParameters', ecsParameters_networkConfiguration - Use this structure if the Amazon ECS task uses the awsvpc network mode. This structure specifies the VPC subnets and security groups associated with the task, and whether a public IP address is to be used. This structure is required if LaunchType is FARGATE because the awsvpc mode is required for Fargate tasks.

If you specify NetworkConfiguration when the target ECS task does not use the awsvpc network mode, the task fails.

$sel:tags:EcsParameters', ecsParameters_tags - The metadata that you apply to the task to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define. To learn more, see RunTask in the Amazon ECS API Reference.

$sel:enableExecuteCommand:EcsParameters', ecsParameters_enableExecuteCommand - Whether or not to enable the execute command functionality for the containers in this task. If true, this enables execute command functionality on all containers in the task.

$sel:taskDefinitionArn:EcsParameters', ecsParameters_taskDefinitionArn - The ARN of the task definition to use if the event target is an Amazon ECS task.

ecsParameters_group :: Lens' EcsParameters (Maybe Text)

Specifies an ECS task group for the task. The maximum length is 255 characters.

ecsParameters_propagateTags :: Lens' EcsParameters (Maybe PropagateTags)

Specifies whether to propagate the tags from the task definition to the task. If no value is specified, the tags are not propagated. Tags can only be propagated to the task during task creation. To add tags to a task after task creation, use the TagResource API action.

ecsParameters_platformVersion :: Lens' EcsParameters (Maybe Text)

Specifies the platform version for the task. Specify only the numeric portion of the platform version, such as 1.1.0.

This structure is used only if LaunchType is FARGATE. For more information about valid platform versions, see Fargate Platform Versions in the Amazon Elastic Container Service Developer Guide.

ecsParameters_enableECSManagedTags :: Lens' EcsParameters (Maybe Bool)

Specifies whether to enable Amazon ECS managed tags for the task. For more information, see Tagging Your Amazon ECS Resources in the Amazon Elastic Container Service Developer Guide.

ecsParameters_referenceId :: Lens' EcsParameters (Maybe Text)

The reference ID to use for the task.

ecsParameters_placementConstraints :: Lens' EcsParameters (Maybe [PlacementConstraint])

An array of placement constraint objects to use for the task. You can specify up to 10 constraints per task (including constraints in the task definition and those specified at runtime).

ecsParameters_placementStrategy :: Lens' EcsParameters (Maybe [PlacementStrategy])

The placement strategy objects to use for the task. You can specify a maximum of five strategy rules per task.

ecsParameters_launchType :: Lens' EcsParameters (Maybe LaunchType)

Specifies the launch type on which your task is running. The launch type that you specify here must match one of the launch type (compatibilities) of the target task. The FARGATE value is supported only in the Regions where Fargate witt Amazon ECS is supported. For more information, see Fargate on Amazon ECS in the Amazon Elastic Container Service Developer Guide.

ecsParameters_capacityProviderStrategy :: Lens' EcsParameters (Maybe [CapacityProviderStrategyItem])

The capacity provider strategy to use for the task.

If a capacityProviderStrategy is specified, the launchType parameter must be omitted. If no capacityProviderStrategy or launchType is specified, the defaultCapacityProviderStrategy for the cluster is used.

ecsParameters_taskCount :: Lens' EcsParameters (Maybe Natural)

The number of tasks to create based on TaskDefinition. The default is 1.

ecsParameters_networkConfiguration :: Lens' EcsParameters (Maybe NetworkConfiguration)

Use this structure if the Amazon ECS task uses the awsvpc network mode. This structure specifies the VPC subnets and security groups associated with the task, and whether a public IP address is to be used. This structure is required if LaunchType is FARGATE because the awsvpc mode is required for Fargate tasks.

If you specify NetworkConfiguration when the target ECS task does not use the awsvpc network mode, the task fails.

ecsParameters_tags :: Lens' EcsParameters (Maybe [Tag])

The metadata that you apply to the task to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define. To learn more, see RunTask in the Amazon ECS API Reference.

ecsParameters_enableExecuteCommand :: Lens' EcsParameters (Maybe Bool)

Whether or not to enable the execute command functionality for the containers in this task. If true, this enables execute command functionality on all containers in the task.

ecsParameters_taskDefinitionArn :: Lens' EcsParameters Text

The ARN of the task definition to use if the event target is an Amazon ECS task.