Network.AWS.CloudWatchEvents.Types.EventSource

Description

Documentation

data EventSource

A partner event source is created by an SaaS partner. If a customer creates a partner event bus that matches this event source, that Amazon Web Services account can receive events from the partner's applications or services.

See: newEventSource smart constructor.

Constructors

EventSource'

Fields

creationTime :: Maybe POSIX

The date and time the event source was created.

state :: Maybe EventSourceState

The state of the event source. If it is ACTIVE, you have already created a matching event bus for this event source, and that event bus is active. If it is PENDING, either you haven't yet created a matching event bus, or that event bus is deactivated. If it is DELETED, you have created a matching event bus, but the event source has since been deleted.

arn :: Maybe Text

The ARN of the event source.

createdBy :: Maybe Text

The name of the partner that created the event source.

name :: Maybe Text

The name of the event source.

expirationTime :: Maybe POSIX

The date and time that the event source will expire, if the Amazon Web Services account doesn't create a matching event bus for it.

newEventSource :: EventSource

Create a value of EventSource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:EventSource', eventSource_creationTime - The date and time the event source was created.

$sel:state:EventSource', eventSource_state - The state of the event source. If it is ACTIVE, you have already created a matching event bus for this event source, and that event bus is active. If it is PENDING, either you haven't yet created a matching event bus, or that event bus is deactivated. If it is DELETED, you have created a matching event bus, but the event source has since been deleted.

$sel:arn:EventSource', eventSource_arn - The ARN of the event source.

$sel:createdBy:EventSource', eventSource_createdBy - The name of the partner that created the event source.

$sel:name:EventSource', eventSource_name - The name of the event source.

$sel:expirationTime:EventSource', eventSource_expirationTime - The date and time that the event source will expire, if the Amazon Web Services account doesn't create a matching event bus for it.

eventSource_creationTime :: Lens' EventSource (Maybe UTCTime)

The date and time the event source was created.

eventSource_state :: Lens' EventSource (Maybe EventSourceState)

The state of the event source. If it is ACTIVE, you have already created a matching event bus for this event source, and that event bus is active. If it is PENDING, either you haven't yet created a matching event bus, or that event bus is deactivated. If it is DELETED, you have created a matching event bus, but the event source has since been deleted.

eventSource_arn :: Lens' EventSource (Maybe Text)

The ARN of the event source.

eventSource_createdBy :: Lens' EventSource (Maybe Text)

The name of the partner that created the event source.

eventSource_name :: Lens' EventSource (Maybe Text)

The name of the event source.

eventSource_expirationTime :: Lens' EventSource (Maybe UTCTime)

The date and time that the event source will expire, if the Amazon Web Services account doesn't create a matching event bus for it.