Network.AWS.CloudWatchEvents.Types.PartnerEventSourceAccount

Description

Documentation

data PartnerEventSourceAccount

The Amazon Web Services account that a partner event source has been offered to.

See: newPartnerEventSourceAccount smart constructor.

Constructors

PartnerEventSourceAccount'

Fields

creationTime :: Maybe POSIX

The date and time the event source was created.

state :: Maybe EventSourceState

The state of the event source. If it is ACTIVE, you have already created a matching event bus for this event source, and that event bus is active. If it is PENDING, either you haven't yet created a matching event bus, or that event bus is deactivated. If it is DELETED, you have created a matching event bus, but the event source has since been deleted.

account :: Maybe Text

The Amazon Web Services account ID that the partner event source was offered to.

expirationTime :: Maybe POSIX

The date and time that the event source will expire, if the Amazon Web Services account doesn't create a matching event bus for it.

newPartnerEventSourceAccount :: PartnerEventSourceAccount

Create a value of PartnerEventSourceAccount with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:PartnerEventSourceAccount', partnerEventSourceAccount_creationTime - The date and time the event source was created.

$sel:state:PartnerEventSourceAccount', partnerEventSourceAccount_state - The state of the event source. If it is ACTIVE, you have already created a matching event bus for this event source, and that event bus is active. If it is PENDING, either you haven't yet created a matching event bus, or that event bus is deactivated. If it is DELETED, you have created a matching event bus, but the event source has since been deleted.

$sel:account:PartnerEventSourceAccount', partnerEventSourceAccount_account - The Amazon Web Services account ID that the partner event source was offered to.

$sel:expirationTime:PartnerEventSourceAccount', partnerEventSourceAccount_expirationTime - The date and time that the event source will expire, if the Amazon Web Services account doesn't create a matching event bus for it.

partnerEventSourceAccount_creationTime :: Lens' PartnerEventSourceAccount (Maybe UTCTime)

The date and time the event source was created.

partnerEventSourceAccount_state :: Lens' PartnerEventSourceAccount (Maybe EventSourceState)

The state of the event source. If it is ACTIVE, you have already created a matching event bus for this event source, and that event bus is active. If it is PENDING, either you haven't yet created a matching event bus, or that event bus is deactivated. If it is DELETED, you have created a matching event bus, but the event source has since been deleted.

partnerEventSourceAccount_account :: Lens' PartnerEventSourceAccount (Maybe Text)

The Amazon Web Services account ID that the partner event source was offered to.

partnerEventSourceAccount_expirationTime :: Lens' PartnerEventSourceAccount (Maybe UTCTime)

The date and time that the event source will expire, if the Amazon Web Services account doesn't create a matching event bus for it.