Network.AWS.CloudWatchEvents.Types.PlacementConstraint

Description

Documentation

data PlacementConstraint

An object representing a constraint on task placement. To learn more, see Task Placement Constraints in the Amazon Elastic Container Service Developer Guide.

See: newPlacementConstraint smart constructor.

Constructors

PlacementConstraint'

Fields

type' :: Maybe PlacementConstraintType

The type of constraint. Use distinctInstance to ensure that each task in a particular group is running on a different container instance. Use memberOf to restrict the selection to a group of valid candidates.

expression :: Maybe Text

A cluster query language expression to apply to the constraint. You cannot specify an expression if the constraint type is distinctInstance. To learn more, see Cluster Query Language in the Amazon Elastic Container Service Developer Guide.

newPlacementConstraint :: PlacementConstraint

Create a value of PlacementConstraint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:type':PlacementConstraint', placementConstraint_type - The type of constraint. Use distinctInstance to ensure that each task in a particular group is running on a different container instance. Use memberOf to restrict the selection to a group of valid candidates.

$sel:expression:PlacementConstraint', placementConstraint_expression - A cluster query language expression to apply to the constraint. You cannot specify an expression if the constraint type is distinctInstance. To learn more, see Cluster Query Language in the Amazon Elastic Container Service Developer Guide.

placementConstraint_type :: Lens' PlacementConstraint (Maybe PlacementConstraintType)

The type of constraint. Use distinctInstance to ensure that each task in a particular group is running on a different container instance. Use memberOf to restrict the selection to a group of valid candidates.

placementConstraint_expression :: Lens' PlacementConstraint (Maybe Text)

A cluster query language expression to apply to the constraint. You cannot specify an expression if the constraint type is distinctInstance. To learn more, see Cluster Query Language in the Amazon Elastic Container Service Developer Guide.