Network.AWS.CloudWatchEvents.Types.PutTargetsResultEntry

Description

Documentation

data PutTargetsResultEntry

Represents a target that failed to be added to a rule.

See: newPutTargetsResultEntry smart constructor.

Constructors

PutTargetsResultEntry'

Fields

targetId :: Maybe Text

The ID of the target.

errorCode :: Maybe Text

The error code that indicates why the target addition failed. If the value is ConcurrentModificationException, too many requests were made at the same time.

errorMessage :: Maybe Text

The error message that explains why the target addition failed.

newPutTargetsResultEntry :: PutTargetsResultEntry

Create a value of PutTargetsResultEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetId:PutTargetsResultEntry', putTargetsResultEntry_targetId - The ID of the target.

$sel:errorCode:PutTargetsResultEntry', putTargetsResultEntry_errorCode - The error code that indicates why the target addition failed. If the value is ConcurrentModificationException, too many requests were made at the same time.

$sel:errorMessage:PutTargetsResultEntry', putTargetsResultEntry_errorMessage - The error message that explains why the target addition failed.

putTargetsResultEntry_targetId :: Lens' PutTargetsResultEntry (Maybe Text)

The ID of the target.

putTargetsResultEntry_errorCode :: Lens' PutTargetsResultEntry (Maybe Text)

The error code that indicates why the target addition failed. If the value is ConcurrentModificationException, too many requests were made at the same time.

putTargetsResultEntry_errorMessage :: Lens' PutTargetsResultEntry (Maybe Text)

The error message that explains why the target addition failed.