Network.AWS.CloudWatchEvents.Types.RedshiftDataParameters

Description

Documentation

data RedshiftDataParameters

These are custom parameters to be used when the target is a Amazon Redshift cluster to invoke the Amazon Redshift Data API ExecuteStatement based on EventBridge events.

See: newRedshiftDataParameters smart constructor.

Constructors

RedshiftDataParameters'

Fields

dbUser :: Maybe Text

The database user name. Required when authenticating using temporary credentials.

secretManagerArn :: Maybe Text

The name or ARN of the secret that enables access to the database. Required when authenticating using Amazon Web Services Secrets Manager.

statementName :: Maybe Text

The name of the SQL statement. You can name the SQL statement when you create it to identify the query.

withEvent :: Maybe Bool

Indicates whether to send an event back to EventBridge after the SQL statement runs.

database :: Text

The name of the database. Required when authenticating using temporary credentials.

sql :: Text

The SQL statement text to run.

newRedshiftDataParameters

Arguments

:: Text
$sel:database:RedshiftDataParameters'

-> Text
$sel:sql:RedshiftDataParameters'

-> RedshiftDataParameters

Create a value of RedshiftDataParameters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbUser:RedshiftDataParameters', redshiftDataParameters_dbUser - The database user name. Required when authenticating using temporary credentials.

$sel:secretManagerArn:RedshiftDataParameters', redshiftDataParameters_secretManagerArn - The name or ARN of the secret that enables access to the database. Required when authenticating using Amazon Web Services Secrets Manager.

$sel:statementName:RedshiftDataParameters', redshiftDataParameters_statementName - The name of the SQL statement. You can name the SQL statement when you create it to identify the query.

$sel:withEvent:RedshiftDataParameters', redshiftDataParameters_withEvent - Indicates whether to send an event back to EventBridge after the SQL statement runs.

$sel:database:RedshiftDataParameters', redshiftDataParameters_database - The name of the database. Required when authenticating using temporary credentials.

$sel:sql:RedshiftDataParameters', redshiftDataParameters_sql - The SQL statement text to run.

redshiftDataParameters_dbUser :: Lens' RedshiftDataParameters (Maybe Text)

The database user name. Required when authenticating using temporary credentials.

redshiftDataParameters_secretManagerArn :: Lens' RedshiftDataParameters (Maybe Text)

The name or ARN of the secret that enables access to the database. Required when authenticating using Amazon Web Services Secrets Manager.

redshiftDataParameters_statementName :: Lens' RedshiftDataParameters (Maybe Text)

The name of the SQL statement. You can name the SQL statement when you create it to identify the query.

redshiftDataParameters_withEvent :: Lens' RedshiftDataParameters (Maybe Bool)

Indicates whether to send an event back to EventBridge after the SQL statement runs.

redshiftDataParameters_database :: Lens' RedshiftDataParameters Text

The name of the database. Required when authenticating using temporary credentials.

redshiftDataParameters_sql :: Lens' RedshiftDataParameters Text

The SQL statement text to run.