Network.AWS.CloudWatchEvents.Types.Rule

Description

Documentation

data Rule

Contains information about a rule in Amazon EventBridge.

See: newRule smart constructor.

Constructors

Rule'

Fields

eventPattern :: Maybe Text

The event pattern of the rule. For more information, see Events and Event Patterns in the Amazon EventBridge User Guide.

roleArn :: Maybe Text

The Amazon Resource Name (ARN) of the role that is used for target invocation.

If you're setting an event bus in another account as the target and that account granted permission to your account through an organization instead of directly by the account ID, you must specify a RoleArn with proper permissions in the Target structure, instead of here in this parameter.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the rule.

name :: Maybe Text

The name of the rule.

eventBusName :: Maybe Text

The name or ARN of the event bus associated with the rule. If you omit this, the default event bus is used.

scheduleExpression :: Maybe Text

The scheduling expression. For example, "cron(0 20 * * ? *)", "rate(5 minutes)". For more information, see Creating an Amazon EventBridge rule that runs on a schedule.

state :: Maybe RuleState

The state of the rule.

managedBy :: Maybe Text

If the rule was created on behalf of your account by an Amazon Web Services service, this field displays the principal name of the service that created the rule.

description :: Maybe Text

The description of the rule.

newRule :: Rule

Create a value of Rule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventPattern:Rule', rule_eventPattern - The event pattern of the rule. For more information, see Events and Event Patterns in the Amazon EventBridge User Guide.

$sel:roleArn:Rule', rule_roleArn - The Amazon Resource Name (ARN) of the role that is used for target invocation.

If you're setting an event bus in another account as the target and that account granted permission to your account through an organization instead of directly by the account ID, you must specify a RoleArn with proper permissions in the Target structure, instead of here in this parameter.

$sel:arn:Rule', rule_arn - The Amazon Resource Name (ARN) of the rule.

$sel:name:Rule', rule_name - The name of the rule.

$sel:eventBusName:Rule', rule_eventBusName - The name or ARN of the event bus associated with the rule. If you omit this, the default event bus is used.

$sel:scheduleExpression:Rule', rule_scheduleExpression - The scheduling expression. For example, "cron(0 20 * * ? *)", "rate(5 minutes)". For more information, see Creating an Amazon EventBridge rule that runs on a schedule.

$sel:state:Rule', rule_state - The state of the rule.

$sel:managedBy:Rule', rule_managedBy - If the rule was created on behalf of your account by an Amazon Web Services service, this field displays the principal name of the service that created the rule.

$sel:description:Rule', rule_description - The description of the rule.

rule_eventPattern :: Lens' Rule (Maybe Text)

The event pattern of the rule. For more information, see Events and Event Patterns in the Amazon EventBridge User Guide.

rule_roleArn :: Lens' Rule (Maybe Text)

The Amazon Resource Name (ARN) of the role that is used for target invocation.

If you're setting an event bus in another account as the target and that account granted permission to your account through an organization instead of directly by the account ID, you must specify a RoleArn with proper permissions in the Target structure, instead of here in this parameter.

rule_arn :: Lens' Rule (Maybe Text)

The Amazon Resource Name (ARN) of the rule.

rule_name :: Lens' Rule (Maybe Text)

The name of the rule.

rule_eventBusName :: Lens' Rule (Maybe Text)

The name or ARN of the event bus associated with the rule. If you omit this, the default event bus is used.

rule_scheduleExpression :: Lens' Rule (Maybe Text)

The scheduling expression. For example, "cron(0 20 * * ? *)", "rate(5 minutes)". For more information, see Creating an Amazon EventBridge rule that runs on a schedule.

rule_state :: Lens' Rule (Maybe RuleState)

The state of the rule.

rule_managedBy :: Lens' Rule (Maybe Text)

If the rule was created on behalf of your account by an Amazon Web Services service, this field displays the principal name of the service that created the rule.

rule_description :: Lens' Rule (Maybe Text)

The description of the rule.