Network.AWS.CloudWatchEvents.Types.RunCommandTarget

Description

Documentation

data RunCommandTarget

Information about the EC2 instances that are to be sent the command, specified as key-value pairs. Each RunCommandTarget block can include only one key, but this key may specify multiple values.

See: newRunCommandTarget smart constructor.

Constructors

RunCommandTarget'

Fields

key :: Text

Can be either tag: tag-key or InstanceIds.

values :: NonEmpty Text

If Key is tag: tag-key, Values is a list of tag values. If Key is InstanceIds, Values is a list of Amazon EC2 instance IDs.

newRunCommandTarget

Arguments

:: Text
$sel:key:RunCommandTarget'

-> NonEmpty Text
$sel:values:RunCommandTarget'

-> RunCommandTarget

Create a value of RunCommandTarget with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:RunCommandTarget', runCommandTarget_key - Can be either tag: tag-key or InstanceIds.

$sel:values:RunCommandTarget', runCommandTarget_values - If Key is tag: tag-key, Values is a list of tag values. If Key is InstanceIds, Values is a list of Amazon EC2 instance IDs.

runCommandTarget_key :: Lens' RunCommandTarget Text

Can be either tag: tag-key or InstanceIds.

runCommandTarget_values :: Lens' RunCommandTarget (NonEmpty Text)

If Key is tag: tag-key, Values is a list of tag values. If Key is InstanceIds, Values is a list of Amazon EC2 instance IDs.