Network.AWS.CloudWatchEvents.Types.Target

Description

Documentation

data Target

Targets are the resources to be invoked when a rule is triggered. For a complete list of services and resources that can be set as a target, see PutTargets.

If you are setting the event bus of another account as the target, and that account granted permission to your account through an organization instead of directly by the account ID, then you must specify a RoleArn with proper permissions in the Target structure. For more information, see Sending and Receiving Events Between Amazon Web Services Accounts in the Amazon EventBridge User Guide.

See: newTarget smart constructor.

Constructors

Target'

Fields

ecsParameters :: Maybe EcsParameters

Contains the Amazon ECS task definition and task count to be used, if the event target is an Amazon ECS task. For more information about Amazon ECS tasks, see Task Definitions in the Amazon EC2 Container Service Developer Guide.

runCommandParameters :: Maybe RunCommandParameters

Parameters used when you are using the rule to invoke Amazon EC2 Run Command.

roleArn :: Maybe Text

The Amazon Resource Name (ARN) of the IAM role to be used for this target when the rule is triggered. If one rule triggers multiple targets, you can use a different IAM role for each target.

redshiftDataParameters :: Maybe RedshiftDataParameters

Contains the Amazon Redshift Data API parameters to use when the target is a Amazon Redshift cluster.

If you specify a Amazon Redshift Cluster as a Target, you can use this to specify parameters to invoke the Amazon Redshift Data API ExecuteStatement based on EventBridge events.

input :: Maybe Text

Valid JSON text passed to the target. In this case, nothing from the event itself is passed to the target. For more information, see The JavaScript Object Notation (JSON) Data Interchange Format.

batchParameters :: Maybe BatchParameters

If the event target is an Batch job, this contains the job definition, job name, and other parameters. For more information, see Jobs in the Batch User Guide.

inputPath :: Maybe Text

The value of the JSONPath that is used for extracting part of the matched event when passing it to the target. You must use JSON dot notation, not bracket notation. For more information about JSON paths, see JSONPath.

retryPolicy :: Maybe RetryPolicy

The RetryPolicy object that contains the retry policy configuration to use for the dead-letter queue.

deadLetterConfig :: Maybe DeadLetterConfig

The DeadLetterConfig that defines the target queue to send dead-letter queue events to.

httpParameters :: Maybe HttpParameters

Contains the HTTP parameters to use when the target is a API Gateway REST endpoint or EventBridge ApiDestination.

If you specify an API Gateway REST API or EventBridge ApiDestination as a target, you can use this parameter to specify headers, path parameters, and query string keys/values as part of your target invoking request. If you're using ApiDestinations, the corresponding Connection can also have these values configured. In case of any conflicting keys, values from the Connection take precedence.

sqsParameters :: Maybe SqsParameters

Contains the message group ID to use when the target is a FIFO queue.

If you specify an SQS FIFO queue as a target, the queue must have content-based deduplication enabled.

sageMakerPipelineParameters :: Maybe SageMakerPipelineParameters

Contains the SageMaker Model Building Pipeline parameters to start execution of a SageMaker Model Building Pipeline.

If you specify a SageMaker Model Building Pipeline as a target, you can use this to specify parameters to start a pipeline execution based on EventBridge events.

inputTransformer :: Maybe InputTransformer

Settings to enable you to provide custom input to a target based on certain event data. You can extract one or more key-value pairs from the event and then use that data to send customized input to the target.

kinesisParameters :: Maybe KinesisParameters

The custom parameter you can use to control the shard assignment, when the target is a Kinesis data stream. If you do not include this parameter, the default is to use the eventId as the partition key.

id :: Text

The ID of the target. We recommend using a memorable and unique string.

arn :: Text

The Amazon Resource Name (ARN) of the target.

newTarget

Arguments

:: Text
$sel:id:Target'

-> Text
$sel:arn:Target'

-> Target

Create a value of Target with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ecsParameters:Target', target_ecsParameters - Contains the Amazon ECS task definition and task count to be used, if the event target is an Amazon ECS task. For more information about Amazon ECS tasks, see Task Definitions in the Amazon EC2 Container Service Developer Guide.

$sel:runCommandParameters:Target', target_runCommandParameters - Parameters used when you are using the rule to invoke Amazon EC2 Run Command.

$sel:roleArn:Target', target_roleArn - The Amazon Resource Name (ARN) of the IAM role to be used for this target when the rule is triggered. If one rule triggers multiple targets, you can use a different IAM role for each target.

$sel:redshiftDataParameters:Target', target_redshiftDataParameters - Contains the Amazon Redshift Data API parameters to use when the target is a Amazon Redshift cluster.

If you specify a Amazon Redshift Cluster as a Target, you can use this to specify parameters to invoke the Amazon Redshift Data API ExecuteStatement based on EventBridge events.

$sel:input:Target', target_input - Valid JSON text passed to the target. In this case, nothing from the event itself is passed to the target. For more information, see The JavaScript Object Notation (JSON) Data Interchange Format.

$sel:batchParameters:Target', target_batchParameters - If the event target is an Batch job, this contains the job definition, job name, and other parameters. For more information, see Jobs in the Batch User Guide.

$sel:inputPath:Target', target_inputPath - The value of the JSONPath that is used for extracting part of the matched event when passing it to the target. You must use JSON dot notation, not bracket notation. For more information about JSON paths, see JSONPath.

$sel:retryPolicy:Target', target_retryPolicy - The RetryPolicy object that contains the retry policy configuration to use for the dead-letter queue.

$sel:deadLetterConfig:Target', target_deadLetterConfig - The DeadLetterConfig that defines the target queue to send dead-letter queue events to.

$sel:httpParameters:Target', target_httpParameters - Contains the HTTP parameters to use when the target is a API Gateway REST endpoint or EventBridge ApiDestination.

If you specify an API Gateway REST API or EventBridge ApiDestination as a target, you can use this parameter to specify headers, path parameters, and query string keys/values as part of your target invoking request. If you're using ApiDestinations, the corresponding Connection can also have these values configured. In case of any conflicting keys, values from the Connection take precedence.

$sel:sqsParameters:Target', target_sqsParameters - Contains the message group ID to use when the target is a FIFO queue.

If you specify an SQS FIFO queue as a target, the queue must have content-based deduplication enabled.

$sel:sageMakerPipelineParameters:Target', target_sageMakerPipelineParameters - Contains the SageMaker Model Building Pipeline parameters to start execution of a SageMaker Model Building Pipeline.

If you specify a SageMaker Model Building Pipeline as a target, you can use this to specify parameters to start a pipeline execution based on EventBridge events.

$sel:inputTransformer:Target', target_inputTransformer - Settings to enable you to provide custom input to a target based on certain event data. You can extract one or more key-value pairs from the event and then use that data to send customized input to the target.

$sel:kinesisParameters:Target', target_kinesisParameters - The custom parameter you can use to control the shard assignment, when the target is a Kinesis data stream. If you do not include this parameter, the default is to use the eventId as the partition key.

$sel:id:Target', target_id - The ID of the target. We recommend using a memorable and unique string.

$sel:arn:Target', target_arn - The Amazon Resource Name (ARN) of the target.

target_ecsParameters :: Lens' Target (Maybe EcsParameters)

Contains the Amazon ECS task definition and task count to be used, if the event target is an Amazon ECS task. For more information about Amazon ECS tasks, see Task Definitions in the Amazon EC2 Container Service Developer Guide.

target_runCommandParameters :: Lens' Target (Maybe RunCommandParameters)

Parameters used when you are using the rule to invoke Amazon EC2 Run Command.

target_roleArn :: Lens' Target (Maybe Text)

The Amazon Resource Name (ARN) of the IAM role to be used for this target when the rule is triggered. If one rule triggers multiple targets, you can use a different IAM role for each target.

target_redshiftDataParameters :: Lens' Target (Maybe RedshiftDataParameters)

Contains the Amazon Redshift Data API parameters to use when the target is a Amazon Redshift cluster.

If you specify a Amazon Redshift Cluster as a Target, you can use this to specify parameters to invoke the Amazon Redshift Data API ExecuteStatement based on EventBridge events.

target_input :: Lens' Target (Maybe Text)

Valid JSON text passed to the target. In this case, nothing from the event itself is passed to the target. For more information, see The JavaScript Object Notation (JSON) Data Interchange Format.

target_batchParameters :: Lens' Target (Maybe BatchParameters)

If the event target is an Batch job, this contains the job definition, job name, and other parameters. For more information, see Jobs in the Batch User Guide.

target_inputPath :: Lens' Target (Maybe Text)

The value of the JSONPath that is used for extracting part of the matched event when passing it to the target. You must use JSON dot notation, not bracket notation. For more information about JSON paths, see JSONPath.

target_retryPolicy :: Lens' Target (Maybe RetryPolicy)

The RetryPolicy object that contains the retry policy configuration to use for the dead-letter queue.

target_deadLetterConfig :: Lens' Target (Maybe DeadLetterConfig)

The DeadLetterConfig that defines the target queue to send dead-letter queue events to.

target_httpParameters :: Lens' Target (Maybe HttpParameters)

Contains the HTTP parameters to use when the target is a API Gateway REST endpoint or EventBridge ApiDestination.

If you specify an API Gateway REST API or EventBridge ApiDestination as a target, you can use this parameter to specify headers, path parameters, and query string keys/values as part of your target invoking request. If you're using ApiDestinations, the corresponding Connection can also have these values configured. In case of any conflicting keys, values from the Connection take precedence.

target_sqsParameters :: Lens' Target (Maybe SqsParameters)

Contains the message group ID to use when the target is a FIFO queue.

If you specify an SQS FIFO queue as a target, the queue must have content-based deduplication enabled.

target_sageMakerPipelineParameters :: Lens' Target (Maybe SageMakerPipelineParameters)

Contains the SageMaker Model Building Pipeline parameters to start execution of a SageMaker Model Building Pipeline.

If you specify a SageMaker Model Building Pipeline as a target, you can use this to specify parameters to start a pipeline execution based on EventBridge events.

target_inputTransformer :: Lens' Target (Maybe InputTransformer)

Settings to enable you to provide custom input to a target based on certain event data. You can extract one or more key-value pairs from the event and then use that data to send customized input to the target.

target_kinesisParameters :: Lens' Target (Maybe KinesisParameters)

The custom parameter you can use to control the shard assignment, when the target is a Kinesis data stream. If you do not include this parameter, the default is to use the eventId as the partition key.

target_id :: Lens' Target Text

The ID of the target. We recommend using a memorable and unique string.

target_arn :: Lens' Target Text

The Amazon Resource Name (ARN) of the target.