Network.AWS.CloudWatchEvents.Types.UpdateConnectionOAuthRequestParameters

Description

Documentation

data UpdateConnectionOAuthRequestParameters

Contains the OAuth request parameters to use for the connection.

See: newUpdateConnectionOAuthRequestParameters smart constructor.

Constructors

UpdateConnectionOAuthRequestParameters'

Fields

httpMethod :: Maybe ConnectionOAuthHttpMethod

The method used to connect to the HTTP endpoint.

clientParameters :: Maybe UpdateConnectionOAuthClientRequestParameters

A UpdateConnectionOAuthClientRequestParameters object that contains the client parameters to use for the connection when OAuth is specified as the authorization type.

authorizationEndpoint :: Maybe Text

The URL to the authorization endpoint when OAuth is specified as the authorization type.

oAuthHttpParameters :: Maybe ConnectionHttpParameters

The additional HTTP parameters used for the OAuth authorization request.

newUpdateConnectionOAuthRequestParameters :: UpdateConnectionOAuthRequestParameters

Create a value of UpdateConnectionOAuthRequestParameters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpMethod:UpdateConnectionOAuthRequestParameters', updateConnectionOAuthRequestParameters_httpMethod - The method used to connect to the HTTP endpoint.

$sel:clientParameters:UpdateConnectionOAuthRequestParameters', updateConnectionOAuthRequestParameters_clientParameters - A UpdateConnectionOAuthClientRequestParameters object that contains the client parameters to use for the connection when OAuth is specified as the authorization type.

$sel:authorizationEndpoint:UpdateConnectionOAuthRequestParameters', updateConnectionOAuthRequestParameters_authorizationEndpoint - The URL to the authorization endpoint when OAuth is specified as the authorization type.

$sel:oAuthHttpParameters:UpdateConnectionOAuthRequestParameters', updateConnectionOAuthRequestParameters_oAuthHttpParameters - The additional HTTP parameters used for the OAuth authorization request.

updateConnectionOAuthRequestParameters_httpMethod :: Lens' UpdateConnectionOAuthRequestParameters (Maybe ConnectionOAuthHttpMethod)

The method used to connect to the HTTP endpoint.

updateConnectionOAuthRequestParameters_clientParameters :: Lens' UpdateConnectionOAuthRequestParameters (Maybe UpdateConnectionOAuthClientRequestParameters)

A UpdateConnectionOAuthClientRequestParameters object that contains the client parameters to use for the connection when OAuth is specified as the authorization type.

updateConnectionOAuthRequestParameters_authorizationEndpoint :: Lens' UpdateConnectionOAuthRequestParameters (Maybe Text)

The URL to the authorization endpoint when OAuth is specified as the authorization type.

updateConnectionOAuthRequestParameters_oAuthHttpParameters :: Lens' UpdateConnectionOAuthRequestParameters (Maybe ConnectionHttpParameters)

The additional HTTP parameters used for the OAuth authorization request.