Network.AWS.CloudWatchEvents.UpdateApiDestination

Description

Updates an API destination.

Creating a Request

data UpdateApiDestination

See: newUpdateApiDestination smart constructor.

Constructors

UpdateApiDestination'

Fields

httpMethod :: Maybe ApiDestinationHttpMethod

The method to use for the API destination.

invocationEndpoint :: Maybe Text

The URL to the endpoint to use for the API destination.

connectionArn :: Maybe Text

The ARN of the connection to use for the API destination.

description :: Maybe Text

The name of the API destination to update.

invocationRateLimitPerSecond :: Maybe Natural

The maximum number of invocations per second to send to the API destination.

name :: Text

The name of the API destination to update.

newUpdateApiDestination

Arguments

:: Text
$sel:name:UpdateApiDestination'

-> UpdateApiDestination

Create a value of UpdateApiDestination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpMethod:UpdateApiDestination', updateApiDestination_httpMethod - The method to use for the API destination.

$sel:invocationEndpoint:UpdateApiDestination', updateApiDestination_invocationEndpoint - The URL to the endpoint to use for the API destination.

$sel:connectionArn:UpdateApiDestination', updateApiDestination_connectionArn - The ARN of the connection to use for the API destination.

$sel:description:UpdateApiDestination', updateApiDestination_description - The name of the API destination to update.

$sel:invocationRateLimitPerSecond:UpdateApiDestination', updateApiDestination_invocationRateLimitPerSecond - The maximum number of invocations per second to send to the API destination.

$sel:name:UpdateApiDestination', updateApiDestination_name - The name of the API destination to update.

Request Lenses

updateApiDestination_httpMethod :: Lens' UpdateApiDestination (Maybe ApiDestinationHttpMethod)

The method to use for the API destination.

updateApiDestination_invocationEndpoint :: Lens' UpdateApiDestination (Maybe Text)

The URL to the endpoint to use for the API destination.

updateApiDestination_connectionArn :: Lens' UpdateApiDestination (Maybe Text)

The ARN of the connection to use for the API destination.

updateApiDestination_description :: Lens' UpdateApiDestination (Maybe Text)

The name of the API destination to update.

updateApiDestination_invocationRateLimitPerSecond :: Lens' UpdateApiDestination (Maybe Natural)

The maximum number of invocations per second to send to the API destination.

updateApiDestination_name :: Lens' UpdateApiDestination Text

The name of the API destination to update.

Destructuring the Response

data UpdateApiDestinationResponse

See: newUpdateApiDestinationResponse smart constructor.

Constructors

UpdateApiDestinationResponse'

Fields

creationTime :: Maybe POSIX

A time stamp for the time that the API destination was created.

apiDestinationArn :: Maybe Text

The ARN of the API destination that was updated.

apiDestinationState :: Maybe ApiDestinationState

The state of the API destination that was updated.

lastModifiedTime :: Maybe POSIX

A time stamp for the time that the API destination was last modified.

httpStatus :: Int

The response's http status code.

newUpdateApiDestinationResponse

Arguments

:: Int
$sel:httpStatus:UpdateApiDestinationResponse'

-> UpdateApiDestinationResponse

Create a value of UpdateApiDestinationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:UpdateApiDestinationResponse', updateApiDestinationResponse_creationTime - A time stamp for the time that the API destination was created.

$sel:apiDestinationArn:UpdateApiDestinationResponse', updateApiDestinationResponse_apiDestinationArn - The ARN of the API destination that was updated.

$sel:apiDestinationState:UpdateApiDestinationResponse', updateApiDestinationResponse_apiDestinationState - The state of the API destination that was updated.

$sel:lastModifiedTime:UpdateApiDestinationResponse', updateApiDestinationResponse_lastModifiedTime - A time stamp for the time that the API destination was last modified.

$sel:httpStatus:UpdateApiDestinationResponse', updateApiDestinationResponse_httpStatus - The response's http status code.

Response Lenses

updateApiDestinationResponse_creationTime :: Lens' UpdateApiDestinationResponse (Maybe UTCTime)

A time stamp for the time that the API destination was created.

updateApiDestinationResponse_apiDestinationArn :: Lens' UpdateApiDestinationResponse (Maybe Text)

The ARN of the API destination that was updated.

updateApiDestinationResponse_apiDestinationState :: Lens' UpdateApiDestinationResponse (Maybe ApiDestinationState)

The state of the API destination that was updated.

updateApiDestinationResponse_lastModifiedTime :: Lens' UpdateApiDestinationResponse (Maybe UTCTime)

A time stamp for the time that the API destination was last modified.

updateApiDestinationResponse_httpStatus :: Lens' UpdateApiDestinationResponse Int

The response's http status code.