Network.AWS.CloudWatchLogs.DeleteQueryDefinition

Description

Deletes a saved CloudWatch Logs Insights query definition. A query definition contains details about a saved CloudWatch Logs Insights query.

Each DeleteQueryDefinition operation can delete one query definition.

You must have the logs:DeleteQueryDefinition permission to be able to perform this operation.

Creating a Request

data DeleteQueryDefinition

See: newDeleteQueryDefinition smart constructor.

Constructors

DeleteQueryDefinition'

Fields

queryDefinitionId :: Text

The ID of the query definition that you want to delete. You can use DescribeQueryDefinitions to retrieve the IDs of your saved query definitions.

newDeleteQueryDefinition

Arguments

:: Text
$sel:queryDefinitionId:DeleteQueryDefinition'

-> DeleteQueryDefinition

Create a value of DeleteQueryDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:queryDefinitionId:DeleteQueryDefinition', deleteQueryDefinition_queryDefinitionId - The ID of the query definition that you want to delete. You can use DescribeQueryDefinitions to retrieve the IDs of your saved query definitions.

Request Lenses

deleteQueryDefinition_queryDefinitionId :: Lens' DeleteQueryDefinition Text

The ID of the query definition that you want to delete. You can use DescribeQueryDefinitions to retrieve the IDs of your saved query definitions.

Destructuring the Response

data DeleteQueryDefinitionResponse

See: newDeleteQueryDefinitionResponse smart constructor.

Constructors

DeleteQueryDefinitionResponse'

Fields

success :: Maybe Bool

A value of TRUE indicates that the operation succeeded. FALSE indicates that the operation failed.

httpStatus :: Int

The response's http status code.

newDeleteQueryDefinitionResponse

Arguments

:: Int
$sel:httpStatus:DeleteQueryDefinitionResponse'

-> DeleteQueryDefinitionResponse

Create a value of DeleteQueryDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:success:DeleteQueryDefinitionResponse', deleteQueryDefinitionResponse_success - A value of TRUE indicates that the operation succeeded. FALSE indicates that the operation failed.

$sel:httpStatus:DeleteQueryDefinitionResponse', deleteQueryDefinitionResponse_httpStatus - The response's http status code.

Response Lenses

deleteQueryDefinitionResponse_success :: Lens' DeleteQueryDefinitionResponse (Maybe Bool)

A value of TRUE indicates that the operation succeeded. FALSE indicates that the operation failed.

deleteQueryDefinitionResponse_httpStatus :: Lens' DeleteQueryDefinitionResponse Int

The response's http status code.