Network.AWS.CloudWatchLogs.DeleteRetentionPolicy

Description

Deletes the specified retention policy.

Log events do not expire if they belong to log groups without a retention policy.

Creating a Request

data DeleteRetentionPolicy

See: newDeleteRetentionPolicy smart constructor.

Constructors

DeleteRetentionPolicy'

Fields

logGroupName :: Text

The name of the log group.

newDeleteRetentionPolicy

Arguments

:: Text
$sel:logGroupName:DeleteRetentionPolicy'

-> DeleteRetentionPolicy

Create a value of DeleteRetentionPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:logGroupName:DeleteRetentionPolicy', deleteRetentionPolicy_logGroupName - The name of the log group.

Request Lenses

deleteRetentionPolicy_logGroupName :: Lens' DeleteRetentionPolicy Text

The name of the log group.

Destructuring the Response

data DeleteRetentionPolicyResponse

See: newDeleteRetentionPolicyResponse smart constructor.

Constructors

DeleteRetentionPolicyResponse'

newDeleteRetentionPolicyResponse :: DeleteRetentionPolicyResponse

Create a value of DeleteRetentionPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.