Network.AWS.CloudWatchLogs.DescribeMetricFilters

Description

Lists the specified metric filters. You can list all of the metric filters or filter the results by log name, prefix, metric name, or metric namespace. The results are ASCII-sorted by filter name.

This operation returns paginated results.

Creating a Request

data DescribeMetricFilters

See: newDescribeMetricFilters smart constructor.

Constructors

DescribeMetricFilters'

Fields

metricNamespace :: Maybe Text

Filters results to include only those in the specified namespace. If you include this parameter in your request, you must also include the metricName parameter.

filterNamePrefix :: Maybe Text

The prefix to match. CloudWatch Logs uses the value you set here only if you also include the logGroupName parameter in your request.

nextToken :: Maybe Text

The token for the next set of items to return. (You received this token from a previous call.)

metricName :: Maybe Text

Filters results to include only those with the specified metric name. If you include this parameter in your request, you must also include the metricNamespace parameter.

logGroupName :: Maybe Text

The name of the log group.

limit :: Maybe Natural

The maximum number of items returned. If you don't specify a value, the default is up to 50 items.

newDescribeMetricFilters :: DescribeMetricFilters

Create a value of DescribeMetricFilters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metricNamespace:DescribeMetricFilters', describeMetricFilters_metricNamespace - Filters results to include only those in the specified namespace. If you include this parameter in your request, you must also include the metricName parameter.

$sel:filterNamePrefix:DescribeMetricFilters', describeMetricFilters_filterNamePrefix - The prefix to match. CloudWatch Logs uses the value you set here only if you also include the logGroupName parameter in your request.

$sel:nextToken:DescribeMetricFilters', describeMetricFilters_nextToken - The token for the next set of items to return. (You received this token from a previous call.)

$sel:metricName:DescribeMetricFilters', describeMetricFilters_metricName - Filters results to include only those with the specified metric name. If you include this parameter in your request, you must also include the metricNamespace parameter.

$sel:logGroupName:DescribeMetricFilters', describeMetricFilters_logGroupName - The name of the log group.

$sel:limit:DescribeMetricFilters', describeMetricFilters_limit - The maximum number of items returned. If you don't specify a value, the default is up to 50 items.

Request Lenses

describeMetricFilters_metricNamespace :: Lens' DescribeMetricFilters (Maybe Text)

Filters results to include only those in the specified namespace. If you include this parameter in your request, you must also include the metricName parameter.

describeMetricFilters_filterNamePrefix :: Lens' DescribeMetricFilters (Maybe Text)

The prefix to match. CloudWatch Logs uses the value you set here only if you also include the logGroupName parameter in your request.

describeMetricFilters_nextToken :: Lens' DescribeMetricFilters (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

describeMetricFilters_metricName :: Lens' DescribeMetricFilters (Maybe Text)

Filters results to include only those with the specified metric name. If you include this parameter in your request, you must also include the metricNamespace parameter.

describeMetricFilters_logGroupName :: Lens' DescribeMetricFilters (Maybe Text)

The name of the log group.

describeMetricFilters_limit :: Lens' DescribeMetricFilters (Maybe Natural)

The maximum number of items returned. If you don't specify a value, the default is up to 50 items.

Destructuring the Response

data DescribeMetricFiltersResponse

See: newDescribeMetricFiltersResponse smart constructor.

Constructors

DescribeMetricFiltersResponse'

Fields

metricFilters :: Maybe [MetricFilter]

The metric filters.

nextToken :: Maybe Text

httpStatus :: Int

The response's http status code.

newDescribeMetricFiltersResponse

Arguments

:: Int
$sel:httpStatus:DescribeMetricFiltersResponse'

-> DescribeMetricFiltersResponse

Create a value of DescribeMetricFiltersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metricFilters:DescribeMetricFiltersResponse', describeMetricFiltersResponse_metricFilters - The metric filters.

$sel:nextToken:DescribeMetricFilters', describeMetricFiltersResponse_nextToken - Undocumented member.

$sel:httpStatus:DescribeMetricFiltersResponse', describeMetricFiltersResponse_httpStatus - The response's http status code.

Response Lenses

describeMetricFiltersResponse_metricFilters :: Lens' DescribeMetricFiltersResponse (Maybe [MetricFilter])

The metric filters.

describeMetricFiltersResponse_nextToken :: Lens' DescribeMetricFiltersResponse (Maybe Text)

Undocumented member.

describeMetricFiltersResponse_httpStatus :: Lens' DescribeMetricFiltersResponse Int

The response's http status code.