Network.AWS.CloudWatchLogs.DisassociateKmsKey

Description

Disassociates the associated Key Management Service customer master key (CMK) from the specified log group.

After the KMS CMK is disassociated from the log group, CloudWatch Logs stops encrypting newly ingested data for the log group. All previously ingested data remains encrypted, and CloudWatch Logs requires permissions for the CMK whenever the encrypted data is requested.

Note that it can take up to 5 minutes for this operation to take effect.

Creating a Request

data DisassociateKmsKey

See: newDisassociateKmsKey smart constructor.

Constructors

DisassociateKmsKey'

Fields

logGroupName :: Text

The name of the log group.

newDisassociateKmsKey

Arguments

:: Text
$sel:logGroupName:DisassociateKmsKey'

-> DisassociateKmsKey

Create a value of DisassociateKmsKey with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:logGroupName:DisassociateKmsKey', disassociateKmsKey_logGroupName - The name of the log group.

Request Lenses

disassociateKmsKey_logGroupName :: Lens' DisassociateKmsKey Text

The name of the log group.

Destructuring the Response

data DisassociateKmsKeyResponse

See: newDisassociateKmsKeyResponse smart constructor.

Constructors

DisassociateKmsKeyResponse'

newDisassociateKmsKeyResponse :: DisassociateKmsKeyResponse

Create a value of DisassociateKmsKeyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.