Network.AWS.CloudWatchLogs.GetLogGroupFields

Description

Returns a list of the fields that are included in log events in the specified log group, along with the percentage of log events that contain each field. The search is limited to a time period that you specify.

In the results, fields that start with @ are fields generated by CloudWatch Logs. For example, @timestamp is the timestamp of each log event. For more information about the fields that are generated by CloudWatch logs, see Supported Logs and Discovered Fields.

The response results are sorted by the frequency percentage, starting with the highest percentage.

Creating a Request

data GetLogGroupFields

See: newGetLogGroupFields smart constructor.

Constructors

GetLogGroupFields'

Fields

time :: Maybe Natural

The time to set as the center of the query. If you specify time, the 15 minutes before this time are queries. If you omit time the 8 minutes before and 8 minutes after this time are searched.

The time value is specified as epoch time, the number of seconds since January 1, 1970, 00:00:00 UTC.

logGroupName :: Text

The name of the log group to search.

newGetLogGroupFields

Arguments

:: Text
$sel:logGroupName:GetLogGroupFields'

-> GetLogGroupFields

Create a value of GetLogGroupFields with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:time:GetLogGroupFields', getLogGroupFields_time - The time to set as the center of the query. If you specify time, the 15 minutes before this time are queries. If you omit time the 8 minutes before and 8 minutes after this time are searched.

The time value is specified as epoch time, the number of seconds since January 1, 1970, 00:00:00 UTC.

$sel:logGroupName:GetLogGroupFields', getLogGroupFields_logGroupName - The name of the log group to search.

Request Lenses

getLogGroupFields_time :: Lens' GetLogGroupFields (Maybe Natural)

The time to set as the center of the query. If you specify time, the 15 minutes before this time are queries. If you omit time the 8 minutes before and 8 minutes after this time are searched.

The time value is specified as epoch time, the number of seconds since January 1, 1970, 00:00:00 UTC.

getLogGroupFields_logGroupName :: Lens' GetLogGroupFields Text

The name of the log group to search.

Destructuring the Response

data GetLogGroupFieldsResponse

See: newGetLogGroupFieldsResponse smart constructor.

Constructors

GetLogGroupFieldsResponse'

Fields

logGroupFields :: Maybe [LogGroupField]

The array of fields found in the query. Each object in the array contains the name of the field, along with the percentage of time it appeared in the log events that were queried.

httpStatus :: Int

The response's http status code.

newGetLogGroupFieldsResponse

Arguments

:: Int
$sel:httpStatus:GetLogGroupFieldsResponse'

-> GetLogGroupFieldsResponse

Create a value of GetLogGroupFieldsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:logGroupFields:GetLogGroupFieldsResponse', getLogGroupFieldsResponse_logGroupFields - The array of fields found in the query. Each object in the array contains the name of the field, along with the percentage of time it appeared in the log events that were queried.

$sel:httpStatus:GetLogGroupFieldsResponse', getLogGroupFieldsResponse_httpStatus - The response's http status code.

Response Lenses

getLogGroupFieldsResponse_logGroupFields :: Lens' GetLogGroupFieldsResponse (Maybe [LogGroupField])

The array of fields found in the query. Each object in the array contains the name of the field, along with the percentage of time it appeared in the log events that were queried.

getLogGroupFieldsResponse_httpStatus :: Lens' GetLogGroupFieldsResponse Int

The response's http status code.