Network.AWS.CloudWatchLogs.Lens

Description

Operations

DescribeExportTasks

describeExportTasks_nextToken :: Lens' DescribeExportTasks (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

describeExportTasks_taskId :: Lens' DescribeExportTasks (Maybe Text)

The ID of the export task. Specifying a task ID filters the results to zero or one export tasks.

describeExportTasks_statusCode :: Lens' DescribeExportTasks (Maybe ExportTaskStatusCode)

The status code of the export task. Specifying a status code filters the results to zero or more export tasks.

describeExportTasks_limit :: Lens' DescribeExportTasks (Maybe Natural)

The maximum number of items returned. If you don't specify a value, the default is up to 50 items.

describeExportTasksResponse_nextToken :: Lens' DescribeExportTasksResponse (Maybe Text)

Undocumented member.

describeExportTasksResponse_exportTasks :: Lens' DescribeExportTasksResponse (Maybe [ExportTask])

The export tasks.

describeExportTasksResponse_httpStatus :: Lens' DescribeExportTasksResponse Int

The response's http status code.

GetLogRecord

getLogRecord_logRecordPointer :: Lens' GetLogRecord Text

The pointer corresponding to the log event record you want to retrieve. You get this from the response of a GetQueryResults operation. In that response, the value of the @ptr field for a log event is the value to use as logRecordPointer to retrieve that complete log event record.

getLogRecordResponse_logRecord :: Lens' GetLogRecordResponse (Maybe (HashMap Text Text))

The requested log event, as a JSON string.

getLogRecordResponse_httpStatus :: Lens' GetLogRecordResponse Int

The response's http status code.

CreateLogStream

createLogStream_logGroupName :: Lens' CreateLogStream Text

The name of the log group.

createLogStream_logStreamName :: Lens' CreateLogStream Text

The name of the log stream.

DescribeResourcePolicies

describeResourcePolicies_nextToken :: Lens' DescribeResourcePolicies (Maybe Text)

Undocumented member.

describeResourcePolicies_limit :: Lens' DescribeResourcePolicies (Maybe Natural)

The maximum number of resource policies to be displayed with one call of this API.

describeResourcePoliciesResponse_nextToken :: Lens' DescribeResourcePoliciesResponse (Maybe Text)

Undocumented member.

describeResourcePoliciesResponse_resourcePolicies :: Lens' DescribeResourcePoliciesResponse (Maybe [ResourcePolicy])

The resource policies that exist in this account.

describeResourcePoliciesResponse_httpStatus :: Lens' DescribeResourcePoliciesResponse Int

The response's http status code.

DescribeQueryDefinitions

describeQueryDefinitions_nextToken :: Lens' DescribeQueryDefinitions (Maybe Text)

Undocumented member.

describeQueryDefinitions_maxResults :: Lens' DescribeQueryDefinitions (Maybe Natural)

Limits the number of returned query definitions to the specified number.

describeQueryDefinitions_queryDefinitionNamePrefix :: Lens' DescribeQueryDefinitions (Maybe Text)

Use this parameter to filter your results to only the query definitions that have names that start with the prefix you specify.

describeQueryDefinitionsResponse_nextToken :: Lens' DescribeQueryDefinitionsResponse (Maybe Text)

Undocumented member.

describeQueryDefinitionsResponse_queryDefinitions :: Lens' DescribeQueryDefinitionsResponse (Maybe [QueryDefinition])

The list of query definitions that match your request.

describeQueryDefinitionsResponse_httpStatus :: Lens' DescribeQueryDefinitionsResponse Int

The response's http status code.

DescribeLogStreams

describeLogStreams_logStreamNamePrefix :: Lens' DescribeLogStreams (Maybe Text)

The prefix to match.

If orderBy is LastEventTime, you cannot specify this parameter.

describeLogStreams_nextToken :: Lens' DescribeLogStreams (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

describeLogStreams_orderBy :: Lens' DescribeLogStreams (Maybe OrderBy)

If the value is LogStreamName, the results are ordered by log stream name. If the value is LastEventTime, the results are ordered by the event time. The default value is LogStreamName.

If you order the results by event time, you cannot specify the logStreamNamePrefix parameter.

lastEventTimestamp represents the time of the most recent log event in the log stream in CloudWatch Logs. This number is expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC. lastEventTimestamp updates on an eventual consistency basis. It typically updates in less than an hour from ingestion, but in rare situations might take longer.

describeLogStreams_descending :: Lens' DescribeLogStreams (Maybe Bool)

If the value is true, results are returned in descending order. If the value is to false, results are returned in ascending order. The default value is false.

describeLogStreams_limit :: Lens' DescribeLogStreams (Maybe Natural)

The maximum number of items returned. If you don't specify a value, the default is up to 50 items.

describeLogStreams_logGroupName :: Lens' DescribeLogStreams Text

The name of the log group.

describeLogStreamsResponse_nextToken :: Lens' DescribeLogStreamsResponse (Maybe Text)

Undocumented member.

describeLogStreamsResponse_logStreams :: Lens' DescribeLogStreamsResponse (Maybe [LogStream])

The log streams.

describeLogStreamsResponse_httpStatus :: Lens' DescribeLogStreamsResponse Int

The response's http status code.

DeleteQueryDefinition

deleteQueryDefinition_queryDefinitionId :: Lens' DeleteQueryDefinition Text

The ID of the query definition that you want to delete. You can use DescribeQueryDefinitions to retrieve the IDs of your saved query definitions.

deleteQueryDefinitionResponse_success :: Lens' DeleteQueryDefinitionResponse (Maybe Bool)

A value of TRUE indicates that the operation succeeded. FALSE indicates that the operation failed.

deleteQueryDefinitionResponse_httpStatus :: Lens' DeleteQueryDefinitionResponse Int

The response's http status code.

PutResourcePolicy

putResourcePolicy_policyDocument :: Lens' PutResourcePolicy (Maybe Text)

Details of the new policy, including the identity of the principal that is enabled to put logs to this account. This is formatted as a JSON string. This parameter is required.

The following example creates a resource policy enabling the Route 53 service to put DNS query logs in to the specified log group. Replace "logArn" with the ARN of your CloudWatch Logs resource, such as a log group or log stream.

CloudWatch Logs also supports aws:SourceArn and aws:SourceAccount condition context keys.

In the example resource policy, you would replace the value of SourceArn with the resource making the call from Route 53 to CloudWatch Logs and replace the value of SourceAccount with the Amazon Web Services account ID making that call.

{ "Version": "2012-10-17", "Statement": [ { "Sid": "Route53LogsToCloudWatchLogs", "Effect": "Allow", "Principal": { "Service": [ "route53.amazonaws.com" ] }, "Action": "logs:PutLogEvents", "Resource": "logArn", "Condition": { "ArnLike": { "aws:SourceArn": "myRoute53ResourceArn" }, "StringEquals": { "aws:SourceAccount": "myAwsAccountId" } } } ] }

putResourcePolicy_policyName :: Lens' PutResourcePolicy (Maybe Text)

Name of the new policy. This parameter is required.

putResourcePolicyResponse_resourcePolicy :: Lens' PutResourcePolicyResponse (Maybe ResourcePolicy)

The new policy.

putResourcePolicyResponse_httpStatus :: Lens' PutResourcePolicyResponse Int

The response's http status code.

DisassociateKmsKey

disassociateKmsKey_logGroupName :: Lens' DisassociateKmsKey Text

The name of the log group.

DeleteDestination

deleteDestination_destinationName :: Lens' DeleteDestination Text

The name of the destination.

DescribeSubscriptionFilters

describeSubscriptionFilters_filterNamePrefix :: Lens' DescribeSubscriptionFilters (Maybe Text)

The prefix to match. If you don't specify a value, no prefix filter is applied.

describeSubscriptionFilters_nextToken :: Lens' DescribeSubscriptionFilters (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

describeSubscriptionFilters_limit :: Lens' DescribeSubscriptionFilters (Maybe Natural)

The maximum number of items returned. If you don't specify a value, the default is up to 50 items.

describeSubscriptionFilters_logGroupName :: Lens' DescribeSubscriptionFilters Text

The name of the log group.

describeSubscriptionFiltersResponse_nextToken :: Lens' DescribeSubscriptionFiltersResponse (Maybe Text)

Undocumented member.

describeSubscriptionFiltersResponse_subscriptionFilters :: Lens' DescribeSubscriptionFiltersResponse (Maybe [SubscriptionFilter])

The subscription filters.

describeSubscriptionFiltersResponse_httpStatus :: Lens' DescribeSubscriptionFiltersResponse Int

The response's http status code.

PutDestinationPolicy

putDestinationPolicy_destinationName :: Lens' PutDestinationPolicy Text

A name for an existing destination.

putDestinationPolicy_accessPolicy :: Lens' PutDestinationPolicy Text

An IAM policy document that authorizes cross-account users to deliver their log events to the associated destination. This can be up to 5120 bytes.

PutRetentionPolicy

putRetentionPolicy_logGroupName :: Lens' PutRetentionPolicy Text

The name of the log group.

putRetentionPolicy_retentionInDays :: Lens' PutRetentionPolicy Int

Undocumented member.

DeleteMetricFilter

deleteMetricFilter_logGroupName :: Lens' DeleteMetricFilter Text

The name of the log group.

deleteMetricFilter_filterName :: Lens' DeleteMetricFilter Text

The name of the metric filter.

DescribeLogGroups

describeLogGroups_nextToken :: Lens' DescribeLogGroups (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

describeLogGroups_logGroupNamePrefix :: Lens' DescribeLogGroups (Maybe Text)

The prefix to match.

describeLogGroups_limit :: Lens' DescribeLogGroups (Maybe Natural)

The maximum number of items returned. If you don't specify a value, the default is up to 50 items.

describeLogGroupsResponse_nextToken :: Lens' DescribeLogGroupsResponse (Maybe Text)

Undocumented member.

describeLogGroupsResponse_logGroups :: Lens' DescribeLogGroupsResponse (Maybe [LogGroup])

The log groups.

If the retentionInDays value if not included for a log group, then that log group is set to have its events never expire.

describeLogGroupsResponse_httpStatus :: Lens' DescribeLogGroupsResponse Int

The response's http status code.

DescribeMetricFilters

describeMetricFilters_metricNamespace :: Lens' DescribeMetricFilters (Maybe Text)

Filters results to include only those in the specified namespace. If you include this parameter in your request, you must also include the metricName parameter.

describeMetricFilters_filterNamePrefix :: Lens' DescribeMetricFilters (Maybe Text)

The prefix to match. CloudWatch Logs uses the value you set here only if you also include the logGroupName parameter in your request.

describeMetricFilters_nextToken :: Lens' DescribeMetricFilters (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

describeMetricFilters_metricName :: Lens' DescribeMetricFilters (Maybe Text)

Filters results to include only those with the specified metric name. If you include this parameter in your request, you must also include the metricNamespace parameter.

describeMetricFilters_logGroupName :: Lens' DescribeMetricFilters (Maybe Text)

The name of the log group.

describeMetricFilters_limit :: Lens' DescribeMetricFilters (Maybe Natural)

The maximum number of items returned. If you don't specify a value, the default is up to 50 items.

describeMetricFiltersResponse_metricFilters :: Lens' DescribeMetricFiltersResponse (Maybe [MetricFilter])

The metric filters.

describeMetricFiltersResponse_nextToken :: Lens' DescribeMetricFiltersResponse (Maybe Text)

Undocumented member.

describeMetricFiltersResponse_httpStatus :: Lens' DescribeMetricFiltersResponse Int

The response's http status code.

PutSubscriptionFilter

putSubscriptionFilter_roleArn :: Lens' PutSubscriptionFilter (Maybe Text)

The ARN of an IAM role that grants CloudWatch Logs permissions to deliver ingested log events to the destination stream. You don't need to provide the ARN when you are working with a logical destination for cross-account delivery.

putSubscriptionFilter_distribution :: Lens' PutSubscriptionFilter (Maybe Distribution)

The method used to distribute log data to the destination. By default, log data is grouped by log stream, but the grouping can be set to random for a more even distribution. This property is only applicable when the destination is an Amazon Kinesis stream.

putSubscriptionFilter_logGroupName :: Lens' PutSubscriptionFilter Text

The name of the log group.

putSubscriptionFilter_filterName :: Lens' PutSubscriptionFilter Text

A name for the subscription filter. If you are updating an existing filter, you must specify the correct name in filterName. To find the name of the filter currently associated with a log group, use DescribeSubscriptionFilters.

putSubscriptionFilter_filterPattern :: Lens' PutSubscriptionFilter Text

A filter pattern for subscribing to a filtered stream of log events.

putSubscriptionFilter_destinationArn :: Lens' PutSubscriptionFilter Text

The ARN of the destination to deliver matching log events to. Currently, the supported destinations are:

An Amazon Kinesis stream belonging to the same account as the subscription filter, for same-account delivery.

A logical destination (specified using an ARN) belonging to a different account, for cross-account delivery.

If you are setting up a cross-account subscription, the destination must have an IAM policy associated with it that allows the sender to send logs to the destination. For more information, see PutDestinationPolicy.

An Amazon Kinesis Firehose delivery stream belonging to the same account as the subscription filter, for same-account delivery.

A Lambda function belonging to the same account as the subscription filter, for same-account delivery.

DescribeDestinations

describeDestinations_nextToken :: Lens' DescribeDestinations (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

describeDestinations_destinationNamePrefix :: Lens' DescribeDestinations (Maybe Text)

The prefix to match. If you don't specify a value, no prefix filter is applied.

describeDestinations_limit :: Lens' DescribeDestinations (Maybe Natural)

The maximum number of items returned. If you don't specify a value, the default is up to 50 items.

describeDestinationsResponse_nextToken :: Lens' DescribeDestinationsResponse (Maybe Text)

Undocumented member.

describeDestinationsResponse_destinations :: Lens' DescribeDestinationsResponse (Maybe [Destination])

The destinations.

describeDestinationsResponse_httpStatus :: Lens' DescribeDestinationsResponse Int

The response's http status code.

CreateLogGroup

createLogGroup_kmsKeyId :: Lens' CreateLogGroup (Maybe Text)

The Amazon Resource Name (ARN) of the CMK to use when encrypting log data. For more information, see Amazon Resource Names - Key Management Service.

createLogGroup_tags :: Lens' CreateLogGroup (Maybe (HashMap Text Text))

The key-value pairs to use for the tags.

CloudWatch Logs doesn’t support IAM policies that prevent users from assigning specified tags to log groups using the aws:Resource/key-name or aws:TagKeys condition keys. For more information about using tags to control access, see Controlling access to Amazon Web Services resources using tags.

createLogGroup_logGroupName :: Lens' CreateLogGroup Text

The name of the log group.

CancelExportTask

cancelExportTask_taskId :: Lens' CancelExportTask Text

The ID of the export task.

GetLogGroupFields

getLogGroupFields_time :: Lens' GetLogGroupFields (Maybe Natural)

The time to set as the center of the query. If you specify time, the 15 minutes before this time are queries. If you omit time the 8 minutes before and 8 minutes after this time are searched.

The time value is specified as epoch time, the number of seconds since January 1, 1970, 00:00:00 UTC.

getLogGroupFields_logGroupName :: Lens' GetLogGroupFields Text

The name of the log group to search.

getLogGroupFieldsResponse_logGroupFields :: Lens' GetLogGroupFieldsResponse (Maybe [LogGroupField])

The array of fields found in the query. Each object in the array contains the name of the field, along with the percentage of time it appeared in the log events that were queried.

getLogGroupFieldsResponse_httpStatus :: Lens' GetLogGroupFieldsResponse Int

The response's http status code.

DeleteLogStream

deleteLogStream_logGroupName :: Lens' DeleteLogStream Text

The name of the log group.

deleteLogStream_logStreamName :: Lens' DeleteLogStream Text

The name of the log stream.

TagLogGroup

tagLogGroup_logGroupName :: Lens' TagLogGroup Text

The name of the log group.

tagLogGroup_tags :: Lens' TagLogGroup (HashMap Text Text)

The key-value pairs to use for the tags.

DescribeQueries

describeQueries_nextToken :: Lens' DescribeQueries (Maybe Text)

Undocumented member.

describeQueries_status :: Lens' DescribeQueries (Maybe QueryStatus)

Limits the returned queries to only those that have the specified status. Valid values are Cancelled, Complete, Failed, Running, and Scheduled.

describeQueries_maxResults :: Lens' DescribeQueries (Maybe Natural)

Limits the number of returned queries to the specified number.

describeQueries_logGroupName :: Lens' DescribeQueries (Maybe Text)

Limits the returned queries to only those for the specified log group.

describeQueriesResponse_nextToken :: Lens' DescribeQueriesResponse (Maybe Text)

Undocumented member.

describeQueriesResponse_queries :: Lens' DescribeQueriesResponse (Maybe [QueryInfo])

The list of queries that match the request.

describeQueriesResponse_httpStatus :: Lens' DescribeQueriesResponse Int

The response's http status code.

AssociateKmsKey

associateKmsKey_logGroupName :: Lens' AssociateKmsKey Text

The name of the log group.

associateKmsKey_kmsKeyId :: Lens' AssociateKmsKey Text

The Amazon Resource Name (ARN) of the CMK to use when encrypting log data. This must be a symmetric CMK. For more information, see Amazon Resource Names - Key Management Service and Using Symmetric and Asymmetric Keys.

GetQueryResults

getQueryResults_queryId :: Lens' GetQueryResults Text

The ID number of the query.

getQueryResultsResponse_status :: Lens' GetQueryResultsResponse (Maybe QueryStatus)

The status of the most recent running of the query. Possible values are Cancelled, Complete, Failed, Running, Scheduled, Timeout, and Unknown.

Queries time out after 15 minutes of execution. To avoid having your queries time out, reduce the time range being searched or partition your query into a number of queries.

getQueryResultsResponse_statistics :: Lens' GetQueryResultsResponse (Maybe QueryStatistics)

Includes the number of log events scanned by the query, the number of log events that matched the query criteria, and the total number of bytes in the log events that were scanned. These values reflect the full raw results of the query.

getQueryResultsResponse_results :: Lens' GetQueryResultsResponse (Maybe [[ResultField]])

The log events that matched the query criteria during the most recent time it ran.

The results value is an array of arrays. Each log event is one object in the top-level array. Each of these log event objects is an array of field/value pairs.

getQueryResultsResponse_httpStatus :: Lens' GetQueryResultsResponse Int

The response's http status code.

PutQueryDefinition

putQueryDefinition_logGroupNames :: Lens' PutQueryDefinition (Maybe [Text])

Use this parameter to include specific log groups as part of your query definition.

If you are updating a query definition and you omit this parameter, then the updated definition will contain no log groups.

putQueryDefinition_queryDefinitionId :: Lens' PutQueryDefinition (Maybe Text)

If you are updating a query definition, use this parameter to specify the ID of the query definition that you want to update. You can use DescribeQueryDefinitions to retrieve the IDs of your saved query definitions.

If you are creating a query definition, do not specify this parameter. CloudWatch generates a unique ID for the new query definition and include it in the response to this operation.

putQueryDefinition_name :: Lens' PutQueryDefinition Text

A name for the query definition. If you are saving a lot of query definitions, we recommend that you name them so that you can easily find the ones you want by using the first part of the name as a filter in the queryDefinitionNamePrefix parameter of DescribeQueryDefinitions.

putQueryDefinition_queryString :: Lens' PutQueryDefinition Text

The query string to use for this definition. For more information, see CloudWatch Logs Insights Query Syntax.

putQueryDefinitionResponse_queryDefinitionId :: Lens' PutQueryDefinitionResponse (Maybe Text)

The ID of the query definition.

putQueryDefinitionResponse_httpStatus :: Lens' PutQueryDefinitionResponse Int

The response's http status code.

DeleteResourcePolicy

deleteResourcePolicy_policyName :: Lens' DeleteResourcePolicy (Maybe Text)

The name of the policy to be revoked. This parameter is required.

FilterLogEvents

filterLogEvents_logStreamNamePrefix :: Lens' FilterLogEvents (Maybe Text)

Filters the results to include only events from log streams that have names starting with this prefix.

If you specify a value for both logStreamNamePrefix and logStreamNames, but the value for logStreamNamePrefix does not match any log stream names specified in logStreamNames, the action returns an InvalidParameterException error.

filterLogEvents_nextToken :: Lens' FilterLogEvents (Maybe Text)

The token for the next set of events to return. (You received this token from a previous call.)

filterLogEvents_interleaved :: Lens' FilterLogEvents (Maybe Bool)

If the value is true, the operation makes a best effort to provide responses that contain events from multiple log streams within the log group, interleaved in a single response. If the value is false, all the matched log events in the first log stream are searched first, then those in the next log stream, and so on. The default is false.

Important: Starting on June 17, 2019, this parameter is ignored and the value is assumed to be true. The response from this operation always interleaves events from multiple log streams within a log group.

filterLogEvents_filterPattern :: Lens' FilterLogEvents (Maybe Text)

The filter pattern to use. For more information, see Filter and Pattern Syntax.

If not provided, all the events are matched.

filterLogEvents_startTime :: Lens' FilterLogEvents (Maybe Natural)

The start of the time range, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC. Events with a timestamp before this time are not returned.

filterLogEvents_endTime :: Lens' FilterLogEvents (Maybe Natural)

The end of the time range, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC. Events with a timestamp later than this time are not returned.

filterLogEvents_logStreamNames :: Lens' FilterLogEvents (Maybe (NonEmpty Text))

Filters the results to only logs from the log streams in this list.

If you specify a value for both logStreamNamePrefix and logStreamNames, the action returns an InvalidParameterException error.

filterLogEvents_limit :: Lens' FilterLogEvents (Maybe Natural)

The maximum number of events to return. The default is 10,000 events.

filterLogEvents_logGroupName :: Lens' FilterLogEvents Text

The name of the log group to search.

filterLogEventsResponse_nextToken :: Lens' FilterLogEventsResponse (Maybe Text)

The token to use when requesting the next set of items. The token expires after 24 hours.

filterLogEventsResponse_searchedLogStreams :: Lens' FilterLogEventsResponse (Maybe [SearchedLogStream])

IMPORTANT Starting on May 15, 2020, this parameter will be deprecated. This parameter will be an empty list after the deprecation occurs.

Indicates which log streams have been searched and whether each has been searched completely.

filterLogEventsResponse_events :: Lens' FilterLogEventsResponse (Maybe [FilteredLogEvent])

The matched events.

filterLogEventsResponse_httpStatus :: Lens' FilterLogEventsResponse Int

The response's http status code.

ListTagsLogGroup

listTagsLogGroup_logGroupName :: Lens' ListTagsLogGroup Text

The name of the log group.

listTagsLogGroupResponse_tags :: Lens' ListTagsLogGroupResponse (Maybe (HashMap Text Text))

The tags for the log group.

listTagsLogGroupResponse_httpStatus :: Lens' ListTagsLogGroupResponse Int

The response's http status code.

GetLogEvents

getLogEvents_nextToken :: Lens' GetLogEvents (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

getLogEvents_startFromHead :: Lens' GetLogEvents (Maybe Bool)

If the value is true, the earliest log events are returned first. If the value is false, the latest log events are returned first. The default value is false.

If you are using a previous nextForwardToken value as the nextToken in this operation, you must specify true for startFromHead.

getLogEvents_startTime :: Lens' GetLogEvents (Maybe Natural)

The start of the time range, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC. Events with a timestamp equal to this time or later than this time are included. Events with a timestamp earlier than this time are not included.

getLogEvents_endTime :: Lens' GetLogEvents (Maybe Natural)

The end of the time range, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC. Events with a timestamp equal to or later than this time are not included.

getLogEvents_limit :: Lens' GetLogEvents (Maybe Natural)

The maximum number of log events returned. If you don't specify a value, the maximum is as many log events as can fit in a response size of 1 MB, up to 10,000 log events.

getLogEvents_logGroupName :: Lens' GetLogEvents Text

The name of the log group.

getLogEvents_logStreamName :: Lens' GetLogEvents Text

The name of the log stream.

getLogEventsResponse_nextForwardToken :: Lens' GetLogEventsResponse (Maybe Text)

The token for the next set of items in the forward direction. The token expires after 24 hours. If you have reached the end of the stream, it returns the same token you passed in.

getLogEventsResponse_nextBackwardToken :: Lens' GetLogEventsResponse (Maybe Text)

The token for the next set of items in the backward direction. The token expires after 24 hours. This token is never null. If you have reached the end of the stream, it returns the same token you passed in.

getLogEventsResponse_events :: Lens' GetLogEventsResponse (Maybe [OutputLogEvent])

The events.

getLogEventsResponse_httpStatus :: Lens' GetLogEventsResponse Int

The response's http status code.

CreateExportTask

createExportTask_logStreamNamePrefix :: Lens' CreateExportTask (Maybe Text)

Export only log streams that match the provided prefix. If you don't specify a value, no prefix filter is applied.

createExportTask_taskName :: Lens' CreateExportTask (Maybe Text)

The name of the export task.

createExportTask_destinationPrefix :: Lens' CreateExportTask (Maybe Text)

The prefix used as the start of the key for every object exported. If you don't specify a value, the default is exportedlogs.

createExportTask_logGroupName :: Lens' CreateExportTask Text

The name of the log group.

createExportTask_from :: Lens' CreateExportTask Natural

The start time of the range for the request, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC. Events with a timestamp earlier than this time are not exported.

createExportTask_to :: Lens' CreateExportTask Natural

The end time of the range for the request, expreswatchlogsdocused as the number of milliseconds after Jan 1, 1970 00:00:00 UTC. Events with a timestamp later than this time are not exported.

createExportTask_destination :: Lens' CreateExportTask Text

The name of S3 bucket for the exported log data. The bucket must be in the same Amazon Web Services region.

createExportTaskResponse_taskId :: Lens' CreateExportTaskResponse (Maybe Text)

The ID of the export task.

createExportTaskResponse_httpStatus :: Lens' CreateExportTaskResponse Int

The response's http status code.

DeleteRetentionPolicy

deleteRetentionPolicy_logGroupName :: Lens' DeleteRetentionPolicy Text

The name of the log group.

PutDestination

putDestination_destinationName :: Lens' PutDestination Text

A name for the destination.

putDestination_targetArn :: Lens' PutDestination Text

The ARN of an Amazon Kinesis stream to which to deliver matching log events.

putDestination_roleArn :: Lens' PutDestination Text

The ARN of an IAM role that grants CloudWatch Logs permissions to call the Amazon Kinesis PutRecord operation on the destination stream.

putDestinationResponse_destination :: Lens' PutDestinationResponse (Maybe Destination)

The destination.

putDestinationResponse_httpStatus :: Lens' PutDestinationResponse Int

The response's http status code.

PutMetricFilter

putMetricFilter_logGroupName :: Lens' PutMetricFilter Text

The name of the log group.

putMetricFilter_filterName :: Lens' PutMetricFilter Text

A name for the metric filter.

putMetricFilter_filterPattern :: Lens' PutMetricFilter Text

A filter pattern for extracting metric data out of ingested log events.

putMetricFilter_metricTransformations :: Lens' PutMetricFilter (NonEmpty MetricTransformation)

A collection of information that defines how metric data gets emitted.

StartQuery

startQuery_logGroupNames :: Lens' StartQuery (Maybe [Text])

The list of log groups to be queried. You can include up to 20 log groups.

A StartQuery operation must include a logGroupNames or a logGroupName parameter, but not both.

startQuery_logGroupName :: Lens' StartQuery (Maybe Text)

The log group on which to perform the query.

A StartQuery operation must include a logGroupNames or a logGroupName parameter, but not both.

startQuery_limit :: Lens' StartQuery (Maybe Natural)

The maximum number of log events to return in the query. If the query string uses the fields command, only the specified fields and their values are returned. The default is 1000.

startQuery_startTime :: Lens' StartQuery Natural

The beginning of the time range to query. The range is inclusive, so the specified start time is included in the query. Specified as epoch time, the number of seconds since January 1, 1970, 00:00:00 UTC.

startQuery_endTime :: Lens' StartQuery Natural

The end of the time range to query. The range is inclusive, so the specified end time is included in the query. Specified as epoch time, the number of seconds since January 1, 1970, 00:00:00 UTC.

startQuery_queryString :: Lens' StartQuery Text

The query string to use. For more information, see CloudWatch Logs Insights Query Syntax.

startQueryResponse_queryId :: Lens' StartQueryResponse (Maybe Text)

The unique ID of the query.

startQueryResponse_httpStatus :: Lens' StartQueryResponse Int

The response's http status code.

PutLogEvents

putLogEvents_sequenceToken :: Lens' PutLogEvents (Maybe Text)

The sequence token obtained from the response of the previous PutLogEvents call. An upload in a newly created log stream does not require a sequence token. You can also get the sequence token using DescribeLogStreams. If you call PutLogEvents twice within a narrow time period using the same value for sequenceToken, both calls might be successful or one might be rejected.

putLogEvents_logGroupName :: Lens' PutLogEvents Text

The name of the log group.

putLogEvents_logStreamName :: Lens' PutLogEvents Text

The name of the log stream.

putLogEvents_logEvents :: Lens' PutLogEvents (NonEmpty InputLogEvent)

The log events.

putLogEventsResponse_nextSequenceToken :: Lens' PutLogEventsResponse (Maybe Text)

The next sequence token.

putLogEventsResponse_rejectedLogEventsInfo :: Lens' PutLogEventsResponse (Maybe RejectedLogEventsInfo)

The rejected events.

putLogEventsResponse_httpStatus :: Lens' PutLogEventsResponse Int

The response's http status code.

DeleteLogGroup

deleteLogGroup_logGroupName :: Lens' DeleteLogGroup Text

The name of the log group.

TestMetricFilter

testMetricFilter_filterPattern :: Lens' TestMetricFilter Text

Undocumented member.

testMetricFilter_logEventMessages :: Lens' TestMetricFilter (NonEmpty Text)

The log event messages to test.

testMetricFilterResponse_matches :: Lens' TestMetricFilterResponse (Maybe [MetricFilterMatchRecord])

The matched events.

testMetricFilterResponse_httpStatus :: Lens' TestMetricFilterResponse Int

The response's http status code.

DeleteSubscriptionFilter

deleteSubscriptionFilter_logGroupName :: Lens' DeleteSubscriptionFilter Text

The name of the log group.

deleteSubscriptionFilter_filterName :: Lens' DeleteSubscriptionFilter Text

The name of the subscription filter.

UntagLogGroup

untagLogGroup_logGroupName :: Lens' UntagLogGroup Text

The name of the log group.

untagLogGroup_tags :: Lens' UntagLogGroup (NonEmpty Text)

The tag keys. The corresponding tags are removed from the log group.

StopQuery

stopQuery_queryId :: Lens' StopQuery Text

The ID number of the query to stop. To find this ID number, use DescribeQueries.

stopQueryResponse_success :: Lens' StopQueryResponse (Maybe Bool)

This is true if the query was stopped by the StopQuery operation.

stopQueryResponse_httpStatus :: Lens' StopQueryResponse Int

The response's http status code.

Types

Destination

destination_creationTime :: Lens' Destination (Maybe Natural)

The creation time of the destination, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

destination_roleArn :: Lens' Destination (Maybe Text)

A role for impersonation, used when delivering log events to the target.

destination_destinationName :: Lens' Destination (Maybe Text)

The name of the destination.

destination_arn :: Lens' Destination (Maybe Text)

The ARN of this destination.

destination_accessPolicy :: Lens' Destination (Maybe Text)

An IAM policy document that governs which Amazon Web Services accounts can create subscription filters against this destination.

destination_targetArn :: Lens' Destination (Maybe Text)

The Amazon Resource Name (ARN) of the physical target where the log events are delivered (for example, a Kinesis stream).

ExportTask

exportTask_executionInfo :: Lens' ExportTask (Maybe ExportTaskExecutionInfo)

Execution information about the export task.

exportTask_status :: Lens' ExportTask (Maybe ExportTaskStatus)

The status of the export task.

exportTask_to :: Lens' ExportTask (Maybe Natural)

The end time, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC. Events with a timestamp later than this time are not exported.

exportTask_taskId :: Lens' ExportTask (Maybe Text)

The ID of the export task.

exportTask_taskName :: Lens' ExportTask (Maybe Text)

The name of the export task.

exportTask_destinationPrefix :: Lens' ExportTask (Maybe Text)

The prefix that was used as the start of Amazon S3 key for every object exported.

exportTask_destination :: Lens' ExportTask (Maybe Text)

The name of the S3 bucket to which the log data was exported.

exportTask_logGroupName :: Lens' ExportTask (Maybe Text)

The name of the log group from which logs data was exported.

exportTask_from :: Lens' ExportTask (Maybe Natural)

The start time, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC. Events with a timestamp before this time are not exported.

ExportTaskExecutionInfo

exportTaskExecutionInfo_creationTime :: Lens' ExportTaskExecutionInfo (Maybe Natural)

The creation time of the export task, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

exportTaskExecutionInfo_completionTime :: Lens' ExportTaskExecutionInfo (Maybe Natural)

The completion time of the export task, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

ExportTaskStatus

exportTaskStatus_message :: Lens' ExportTaskStatus (Maybe Text)

The status message related to the status code.

exportTaskStatus_code :: Lens' ExportTaskStatus (Maybe ExportTaskStatusCode)

The status code of the export task.

FilteredLogEvent

filteredLogEvent_logStreamName :: Lens' FilteredLogEvent (Maybe Text)

The name of the log stream to which this event belongs.

filteredLogEvent_eventId :: Lens' FilteredLogEvent (Maybe Text)

The ID of the event.

filteredLogEvent_message :: Lens' FilteredLogEvent (Maybe Text)

The data contained in the log event.

filteredLogEvent_ingestionTime :: Lens' FilteredLogEvent (Maybe Natural)

The time the event was ingested, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

filteredLogEvent_timestamp :: Lens' FilteredLogEvent (Maybe Natural)

The time the event occurred, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

InputLogEvent

inputLogEvent_timestamp :: Lens' InputLogEvent Natural

The time the event occurred, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

inputLogEvent_message :: Lens' InputLogEvent Text

The raw event message.

LogGroup

logGroup_retentionInDays :: Lens' LogGroup (Maybe Int)

Undocumented member.

logGroup_creationTime :: Lens' LogGroup (Maybe Natural)

The creation time of the log group, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

logGroup_arn :: Lens' LogGroup (Maybe Text)

The Amazon Resource Name (ARN) of the log group.

logGroup_storedBytes :: Lens' LogGroup (Maybe Natural)

The number of bytes stored.

logGroup_kmsKeyId :: Lens' LogGroup (Maybe Text)

The Amazon Resource Name (ARN) of the CMK to use when encrypting log data.

logGroup_metricFilterCount :: Lens' LogGroup (Maybe Int)

The number of metric filters.

logGroup_logGroupName :: Lens' LogGroup (Maybe Text)

The name of the log group.

LogGroupField

logGroupField_name :: Lens' LogGroupField (Maybe Text)

The name of a log field.

logGroupField_percent :: Lens' LogGroupField (Maybe Natural)

The percentage of log events queried that contained the field.

LogStream

logStream_logStreamName :: Lens' LogStream (Maybe Text)

The name of the log stream.

logStream_creationTime :: Lens' LogStream (Maybe Natural)

The creation time of the stream, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

logStream_arn :: Lens' LogStream (Maybe Text)

The Amazon Resource Name (ARN) of the log stream.

logStream_storedBytes :: Lens' LogStream (Maybe Natural)

The number of bytes stored.

Important: On June 17, 2019, this parameter was deprecated for log streams, and is always reported as zero. This change applies only to log streams. The storedBytes parameter for log groups is not affected.

logStream_uploadSequenceToken :: Lens' LogStream (Maybe Text)

The sequence token.

logStream_lastEventTimestamp :: Lens' LogStream (Maybe Natural)

The time of the most recent log event in the log stream in CloudWatch Logs. This number is expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC. The lastEventTime value updates on an eventual consistency basis. It typically updates in less than an hour from ingestion, but in rare situations might take longer.

logStream_firstEventTimestamp :: Lens' LogStream (Maybe Natural)

The time of the first event, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

logStream_lastIngestionTime :: Lens' LogStream (Maybe Natural)

The ingestion time, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

MetricFilter

metricFilter_filterName :: Lens' MetricFilter (Maybe Text)

The name of the metric filter.

metricFilter_creationTime :: Lens' MetricFilter (Maybe Natural)

The creation time of the metric filter, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

metricFilter_filterPattern :: Lens' MetricFilter (Maybe Text)

Undocumented member.

metricFilter_logGroupName :: Lens' MetricFilter (Maybe Text)

The name of the log group.

metricFilter_metricTransformations :: Lens' MetricFilter (Maybe (NonEmpty MetricTransformation))

The metric transformations.

MetricFilterMatchRecord

metricFilterMatchRecord_eventNumber :: Lens' MetricFilterMatchRecord (Maybe Integer)

The event number.

metricFilterMatchRecord_eventMessage :: Lens' MetricFilterMatchRecord (Maybe Text)

The raw event data.

metricFilterMatchRecord_extractedValues :: Lens' MetricFilterMatchRecord (Maybe (HashMap Text Text))

The values extracted from the event data by the filter.

MetricTransformation

metricTransformation_unit :: Lens' MetricTransformation (Maybe StandardUnit)

The unit to assign to the metric. If you omit this, the unit is set as None.

metricTransformation_dimensions :: Lens' MetricTransformation (Maybe (HashMap Text Text))

The fields to use as dimensions for the metric. One metric filter can include as many as three dimensions.

Metrics extracted from log events are charged as custom metrics. To prevent unexpected high charges, do not specify high-cardinality fields such as IPAddress or requestID as dimensions. Each different value found for a dimension is treated as a separate metric and accrues charges as a separate custom metric.

To help prevent accidental high charges, Amazon disables a metric filter if it generates 1000 different name/value pairs for the dimensions that you have specified within a certain amount of time.

You can also set up a billing alarm to alert you if your charges are higher than expected. For more information, see Creating a Billing Alarm to Monitor Your Estimated Amazon Web Services Charges.

metricTransformation_defaultValue :: Lens' MetricTransformation (Maybe Double)

(Optional) The value to emit when a filter pattern does not match a log event. This value can be null.

metricTransformation_metricName :: Lens' MetricTransformation Text

The name of the CloudWatch metric.

metricTransformation_metricNamespace :: Lens' MetricTransformation Text

A custom namespace to contain your metric in CloudWatch. Use namespaces to group together metrics that are similar. For more information, see Namespaces.

metricTransformation_metricValue :: Lens' MetricTransformation Text

The value to publish to the CloudWatch metric when a filter pattern matches a log event.

OutputLogEvent

outputLogEvent_message :: Lens' OutputLogEvent (Maybe Text)

The data contained in the log event.

outputLogEvent_ingestionTime :: Lens' OutputLogEvent (Maybe Natural)

The time the event was ingested, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

outputLogEvent_timestamp :: Lens' OutputLogEvent (Maybe Natural)

The time the event occurred, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

QueryDefinition

queryDefinition_queryString :: Lens' QueryDefinition (Maybe Text)

The query string to use for this definition. For more information, see CloudWatch Logs Insights Query Syntax.

queryDefinition_name :: Lens' QueryDefinition (Maybe Text)

The name of the query definition.

queryDefinition_logGroupNames :: Lens' QueryDefinition (Maybe [Text])

If this query definition contains a list of log groups that it is limited to, that list appears here.

queryDefinition_lastModified :: Lens' QueryDefinition (Maybe Natural)

The date that the query definition was most recently modified.

queryDefinition_queryDefinitionId :: Lens' QueryDefinition (Maybe Text)

The unique ID of the query definition.

QueryInfo

queryInfo_queryString :: Lens' QueryInfo (Maybe Text)

The query string used in this query.

queryInfo_status :: Lens' QueryInfo (Maybe QueryStatus)

The status of this query. Possible values are Cancelled, Complete, Failed, Running, Scheduled, and Unknown.

queryInfo_queryId :: Lens' QueryInfo (Maybe Text)

The unique ID number of this query.

queryInfo_logGroupName :: Lens' QueryInfo (Maybe Text)

The name of the log group scanned by this query.

queryInfo_createTime :: Lens' QueryInfo (Maybe Natural)

The date and time that this query was created.

QueryStatistics

queryStatistics_bytesScanned :: Lens' QueryStatistics (Maybe Double)

The total number of bytes in the log events scanned during the query.

queryStatistics_recordsMatched :: Lens' QueryStatistics (Maybe Double)

The number of log events that matched the query string.

queryStatistics_recordsScanned :: Lens' QueryStatistics (Maybe Double)

The total number of log events scanned during the query.

RejectedLogEventsInfo

rejectedLogEventsInfo_tooOldLogEventEndIndex :: Lens' RejectedLogEventsInfo (Maybe Int)

The log events that are too old.

rejectedLogEventsInfo_expiredLogEventEndIndex :: Lens' RejectedLogEventsInfo (Maybe Int)

The expired log events.

rejectedLogEventsInfo_tooNewLogEventStartIndex :: Lens' RejectedLogEventsInfo (Maybe Int)

The log events that are too new.

ResourcePolicy

resourcePolicy_policyDocument :: Lens' ResourcePolicy (Maybe Text)

The details of the policy.

resourcePolicy_policyName :: Lens' ResourcePolicy (Maybe Text)

The name of the resource policy.

resourcePolicy_lastUpdatedTime :: Lens' ResourcePolicy (Maybe Natural)

Timestamp showing when this policy was last updated, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

ResultField

resultField_value :: Lens' ResultField (Maybe Text)

The value of this field.

resultField_field :: Lens' ResultField (Maybe Text)

The log event field.

SearchedLogStream

searchedLogStream_logStreamName :: Lens' SearchedLogStream (Maybe Text)

The name of the log stream.

searchedLogStream_searchedCompletely :: Lens' SearchedLogStream (Maybe Bool)

Indicates whether all the events in this log stream were searched.

SubscriptionFilter

subscriptionFilter_filterName :: Lens' SubscriptionFilter (Maybe Text)

The name of the subscription filter.

subscriptionFilter_creationTime :: Lens' SubscriptionFilter (Maybe Natural)

The creation time of the subscription filter, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

subscriptionFilter_destinationArn :: Lens' SubscriptionFilter (Maybe Text)

The Amazon Resource Name (ARN) of the destination.

subscriptionFilter_roleArn :: Lens' SubscriptionFilter (Maybe Text)

subscriptionFilter_filterPattern :: Lens' SubscriptionFilter (Maybe Text)

Undocumented member.

subscriptionFilter_distribution :: Lens' SubscriptionFilter (Maybe Distribution)

Undocumented member.

subscriptionFilter_logGroupName :: Lens' SubscriptionFilter (Maybe Text)

The name of the log group.