Network.AWS.CloudWatchLogs.ListTagsLogGroup

Description

Lists the tags for the specified log group.

Creating a Request

data ListTagsLogGroup

See: newListTagsLogGroup smart constructor.

Constructors

ListTagsLogGroup'

Fields

logGroupName :: Text

The name of the log group.

newListTagsLogGroup

Arguments

:: Text
$sel:logGroupName:ListTagsLogGroup'

-> ListTagsLogGroup

Create a value of ListTagsLogGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:logGroupName:ListTagsLogGroup', listTagsLogGroup_logGroupName - The name of the log group.

Request Lenses

listTagsLogGroup_logGroupName :: Lens' ListTagsLogGroup Text

The name of the log group.

Destructuring the Response

data ListTagsLogGroupResponse

See: newListTagsLogGroupResponse smart constructor.

Constructors

ListTagsLogGroupResponse'

Fields

tags :: Maybe (HashMap Text Text)

The tags for the log group.

httpStatus :: Int

The response's http status code.

newListTagsLogGroupResponse

Arguments

:: Int
$sel:httpStatus:ListTagsLogGroupResponse'

-> ListTagsLogGroupResponse

Create a value of ListTagsLogGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:ListTagsLogGroupResponse', listTagsLogGroupResponse_tags - The tags for the log group.

$sel:httpStatus:ListTagsLogGroupResponse', listTagsLogGroupResponse_httpStatus - The response's http status code.

Response Lenses

listTagsLogGroupResponse_tags :: Lens' ListTagsLogGroupResponse (Maybe (HashMap Text Text))

The tags for the log group.

listTagsLogGroupResponse_httpStatus :: Lens' ListTagsLogGroupResponse Int

The response's http status code.