Network.AWS.CloudWatchLogs.PutLogEvents

Description

Uploads a batch of log events to the specified log stream.

You must include the sequence token obtained from the response of the previous call. An upload in a newly created log stream does not require a sequence token. You can also get the sequence token in the expectedSequenceToken field from InvalidSequenceTokenException. If you call PutLogEvents twice within a narrow time period using the same value for sequenceToken, both calls might be successful or one might be rejected.

The batch of events must satisfy the following constraints:

The maximum batch size is 1,048,576 bytes. This size is calculated as the sum of all event messages in UTF-8, plus 26 bytes for each log event.

None of the log events in the batch can be more than 2 hours in the future.

None of the log events in the batch can be older than 14 days or older than the retention period of the log group.

The log events in the batch must be in chronological order by their timestamp. The timestamp is the time the event occurred, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC. (In Amazon Web Services Tools for PowerShell and the Amazon Web Services SDK for .NET, the timestamp is specified in .NET format: yyyy-mm-ddThh:mm:ss. For example, 2017-09-15T13:45:30.)

A batch of log events in a single request cannot span more than 24 hours. Otherwise, the operation fails.

The maximum number of log events in a batch is 10,000.

There is a quota of 5 requests per second per log stream. Additional requests are throttled. This quota can't be changed.

If a call to PutLogEvents returns "UnrecognizedClientException" the most likely cause is an invalid Amazon Web Services access key ID or secret key.

Creating a Request

data PutLogEvents

See: newPutLogEvents smart constructor.

Constructors

PutLogEvents'

Fields

sequenceToken :: Maybe Text

The sequence token obtained from the response of the previous PutLogEvents call. An upload in a newly created log stream does not require a sequence token. You can also get the sequence token using DescribeLogStreams. If you call PutLogEvents twice within a narrow time period using the same value for sequenceToken, both calls might be successful or one might be rejected.

logGroupName :: Text

The name of the log group.

logStreamName :: Text

The name of the log stream.

logEvents :: NonEmpty InputLogEvent

The log events.

newPutLogEvents

Arguments

:: Text
$sel:logGroupName:PutLogEvents'

-> Text
$sel:logStreamName:PutLogEvents'

-> NonEmpty InputLogEvent
$sel:logEvents:PutLogEvents'

-> PutLogEvents

Create a value of PutLogEvents with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sequenceToken:PutLogEvents', putLogEvents_sequenceToken - The sequence token obtained from the response of the previous PutLogEvents call. An upload in a newly created log stream does not require a sequence token. You can also get the sequence token using DescribeLogStreams. If you call PutLogEvents twice within a narrow time period using the same value for sequenceToken, both calls might be successful or one might be rejected.

$sel:logGroupName:PutLogEvents', putLogEvents_logGroupName - The name of the log group.

$sel:logStreamName:PutLogEvents', putLogEvents_logStreamName - The name of the log stream.

$sel:logEvents:PutLogEvents', putLogEvents_logEvents - The log events.

Request Lenses

putLogEvents_sequenceToken :: Lens' PutLogEvents (Maybe Text)

The sequence token obtained from the response of the previous PutLogEvents call. An upload in a newly created log stream does not require a sequence token. You can also get the sequence token using DescribeLogStreams. If you call PutLogEvents twice within a narrow time period using the same value for sequenceToken, both calls might be successful or one might be rejected.

putLogEvents_logGroupName :: Lens' PutLogEvents Text

The name of the log group.

putLogEvents_logStreamName :: Lens' PutLogEvents Text

The name of the log stream.

putLogEvents_logEvents :: Lens' PutLogEvents (NonEmpty InputLogEvent)

The log events.

Destructuring the Response

data PutLogEventsResponse

See: newPutLogEventsResponse smart constructor.

Constructors

PutLogEventsResponse'

Fields

nextSequenceToken :: Maybe Text

The next sequence token.

rejectedLogEventsInfo :: Maybe RejectedLogEventsInfo

The rejected events.

httpStatus :: Int

The response's http status code.

newPutLogEventsResponse

Arguments

:: Int
$sel:httpStatus:PutLogEventsResponse'

-> PutLogEventsResponse

Create a value of PutLogEventsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextSequenceToken:PutLogEventsResponse', putLogEventsResponse_nextSequenceToken - The next sequence token.

$sel:rejectedLogEventsInfo:PutLogEventsResponse', putLogEventsResponse_rejectedLogEventsInfo - The rejected events.

$sel:httpStatus:PutLogEventsResponse', putLogEventsResponse_httpStatus - The response's http status code.

Response Lenses

putLogEventsResponse_nextSequenceToken :: Lens' PutLogEventsResponse (Maybe Text)

The next sequence token.

putLogEventsResponse_rejectedLogEventsInfo :: Lens' PutLogEventsResponse (Maybe RejectedLogEventsInfo)

The rejected events.

putLogEventsResponse_httpStatus :: Lens' PutLogEventsResponse Int

The response's http status code.