Network.AWS.CloudWatchLogs.PutSubscriptionFilter

Description

Creates or updates a subscription filter and associates it with the specified log group. Subscription filters allow you to subscribe to a real-time stream of log events ingested through PutLogEvents and have them delivered to a specific destination. When log events are sent to the receiving service, they are Base64 encoded and compressed with the gzip format.

The following destinations are supported for subscription filters:

An Amazon Kinesis stream belonging to the same account as the subscription filter, for same-account delivery.

A logical destination that belongs to a different account, for cross-account delivery.

An Amazon Kinesis Firehose delivery stream that belongs to the same account as the subscription filter, for same-account delivery.

An Lambda function that belongs to the same account as the subscription filter, for same-account delivery.

Each log group can have up to two subscription filters associated with it. If you are updating an existing filter, you must specify the correct name in filterName.

To perform a PutSubscriptionFilter operation, you must also have the iam:PassRole permission.

Creating a Request

data PutSubscriptionFilter

See: newPutSubscriptionFilter smart constructor.

Constructors

PutSubscriptionFilter'

Fields

distribution :: Maybe Distribution

The method used to distribute log data to the destination. By default, log data is grouped by log stream, but the grouping can be set to random for a more even distribution. This property is only applicable when the destination is an Amazon Kinesis stream.

roleArn :: Maybe Text

The ARN of an IAM role that grants CloudWatch Logs permissions to deliver ingested log events to the destination stream. You don't need to provide the ARN when you are working with a logical destination for cross-account delivery.

logGroupName :: Text

The name of the log group.

filterName :: Text

A name for the subscription filter. If you are updating an existing filter, you must specify the correct name in filterName. To find the name of the filter currently associated with a log group, use DescribeSubscriptionFilters.

filterPattern :: Text

A filter pattern for subscribing to a filtered stream of log events.

destinationArn :: Text

The ARN of the destination to deliver matching log events to. Currently, the supported destinations are:

An Amazon Kinesis stream belonging to the same account as the subscription filter, for same-account delivery.

A logical destination (specified using an ARN) belonging to a different account, for cross-account delivery.

If you are setting up a cross-account subscription, the destination must have an IAM policy associated with it that allows the sender to send logs to the destination. For more information, see PutDestinationPolicy.

An Amazon Kinesis Firehose delivery stream belonging to the same account as the subscription filter, for same-account delivery.

A Lambda function belonging to the same account as the subscription filter, for same-account delivery.

newPutSubscriptionFilter

Arguments

:: Text
$sel:logGroupName:PutSubscriptionFilter'

-> Text
$sel:filterName:PutSubscriptionFilter'

-> Text
$sel:filterPattern:PutSubscriptionFilter'

-> Text
$sel:destinationArn:PutSubscriptionFilter'

-> PutSubscriptionFilter

Create a value of PutSubscriptionFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:distribution:PutSubscriptionFilter', putSubscriptionFilter_distribution - The method used to distribute log data to the destination. By default, log data is grouped by log stream, but the grouping can be set to random for a more even distribution. This property is only applicable when the destination is an Amazon Kinesis stream.

$sel:roleArn:PutSubscriptionFilter', putSubscriptionFilter_roleArn - The ARN of an IAM role that grants CloudWatch Logs permissions to deliver ingested log events to the destination stream. You don't need to provide the ARN when you are working with a logical destination for cross-account delivery.

$sel:logGroupName:PutSubscriptionFilter', putSubscriptionFilter_logGroupName - The name of the log group.

$sel:filterName:PutSubscriptionFilter', putSubscriptionFilter_filterName - A name for the subscription filter. If you are updating an existing filter, you must specify the correct name in filterName. To find the name of the filter currently associated with a log group, use DescribeSubscriptionFilters.

$sel:filterPattern:PutSubscriptionFilter', putSubscriptionFilter_filterPattern - A filter pattern for subscribing to a filtered stream of log events.

$sel:destinationArn:PutSubscriptionFilter', putSubscriptionFilter_destinationArn - The ARN of the destination to deliver matching log events to. Currently, the supported destinations are:

An Amazon Kinesis stream belonging to the same account as the subscription filter, for same-account delivery.

A logical destination (specified using an ARN) belonging to a different account, for cross-account delivery.

If you are setting up a cross-account subscription, the destination must have an IAM policy associated with it that allows the sender to send logs to the destination. For more information, see PutDestinationPolicy.

An Amazon Kinesis Firehose delivery stream belonging to the same account as the subscription filter, for same-account delivery.

A Lambda function belonging to the same account as the subscription filter, for same-account delivery.

Request Lenses

putSubscriptionFilter_distribution :: Lens' PutSubscriptionFilter (Maybe Distribution)

The method used to distribute log data to the destination. By default, log data is grouped by log stream, but the grouping can be set to random for a more even distribution. This property is only applicable when the destination is an Amazon Kinesis stream.

putSubscriptionFilter_roleArn :: Lens' PutSubscriptionFilter (Maybe Text)

The ARN of an IAM role that grants CloudWatch Logs permissions to deliver ingested log events to the destination stream. You don't need to provide the ARN when you are working with a logical destination for cross-account delivery.

putSubscriptionFilter_logGroupName :: Lens' PutSubscriptionFilter Text

The name of the log group.

putSubscriptionFilter_filterName :: Lens' PutSubscriptionFilter Text

A name for the subscription filter. If you are updating an existing filter, you must specify the correct name in filterName. To find the name of the filter currently associated with a log group, use DescribeSubscriptionFilters.

putSubscriptionFilter_filterPattern :: Lens' PutSubscriptionFilter Text

A filter pattern for subscribing to a filtered stream of log events.

putSubscriptionFilter_destinationArn :: Lens' PutSubscriptionFilter Text

The ARN of the destination to deliver matching log events to. Currently, the supported destinations are:

An Amazon Kinesis stream belonging to the same account as the subscription filter, for same-account delivery.

A logical destination (specified using an ARN) belonging to a different account, for cross-account delivery.

If you are setting up a cross-account subscription, the destination must have an IAM policy associated with it that allows the sender to send logs to the destination. For more information, see PutDestinationPolicy.

An Amazon Kinesis Firehose delivery stream belonging to the same account as the subscription filter, for same-account delivery.

A Lambda function belonging to the same account as the subscription filter, for same-account delivery.

Destructuring the Response

data PutSubscriptionFilterResponse

See: newPutSubscriptionFilterResponse smart constructor.

Constructors

PutSubscriptionFilterResponse'

newPutSubscriptionFilterResponse :: PutSubscriptionFilterResponse

Create a value of PutSubscriptionFilterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.