Network.AWS.CloudWatchLogs.Types.Destination

Description

Documentation

data Destination

Represents a cross-account destination that receives subscription log events.

See: newDestination smart constructor.

Constructors

Destination'

Fields

targetArn :: Maybe Text

The Amazon Resource Name (ARN) of the physical target where the log events are delivered (for example, a Kinesis stream).

creationTime :: Maybe Natural

The creation time of the destination, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

arn :: Maybe Text

The ARN of this destination.

accessPolicy :: Maybe Text

An IAM policy document that governs which Amazon Web Services accounts can create subscription filters against this destination.

destinationName :: Maybe Text

The name of the destination.

roleArn :: Maybe Text

A role for impersonation, used when delivering log events to the target.

newDestination :: Destination

Create a value of Destination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetArn:Destination', destination_targetArn - The Amazon Resource Name (ARN) of the physical target where the log events are delivered (for example, a Kinesis stream).

$sel:creationTime:Destination', destination_creationTime - The creation time of the destination, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

$sel:arn:Destination', destination_arn - The ARN of this destination.

$sel:accessPolicy:Destination', destination_accessPolicy - An IAM policy document that governs which Amazon Web Services accounts can create subscription filters against this destination.

$sel:destinationName:Destination', destination_destinationName - The name of the destination.

$sel:roleArn:Destination', destination_roleArn - A role for impersonation, used when delivering log events to the target.

destination_targetArn :: Lens' Destination (Maybe Text)

The Amazon Resource Name (ARN) of the physical target where the log events are delivered (for example, a Kinesis stream).

destination_creationTime :: Lens' Destination (Maybe Natural)

The creation time of the destination, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

destination_arn :: Lens' Destination (Maybe Text)

The ARN of this destination.

destination_accessPolicy :: Lens' Destination (Maybe Text)

An IAM policy document that governs which Amazon Web Services accounts can create subscription filters against this destination.

destination_destinationName :: Lens' Destination (Maybe Text)

The name of the destination.

destination_roleArn :: Lens' Destination (Maybe Text)

A role for impersonation, used when delivering log events to the target.