Network.AWS.CloudWatchLogs.Types.LogStream

Description

Documentation

data LogStream

Represents a log stream, which is a sequence of log events from a single emitter of logs.

See: newLogStream smart constructor.

Constructors

LogStream'

Fields

logStreamName :: Maybe Text

The name of the log stream.

creationTime :: Maybe Natural

The creation time of the stream, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the log stream.

storedBytes :: Maybe Natural

The number of bytes stored.

Important: On June 17, 2019, this parameter was deprecated for log streams, and is always reported as zero. This change applies only to log streams. The storedBytes parameter for log groups is not affected.

uploadSequenceToken :: Maybe Text

The sequence token.

lastEventTimestamp :: Maybe Natural

The time of the most recent log event in the log stream in CloudWatch Logs. This number is expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC. The lastEventTime value updates on an eventual consistency basis. It typically updates in less than an hour from ingestion, but in rare situations might take longer.

firstEventTimestamp :: Maybe Natural

The time of the first event, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

lastIngestionTime :: Maybe Natural

The ingestion time, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

newLogStream :: LogStream

Create a value of LogStream with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:logStreamName:LogStream', logStream_logStreamName - The name of the log stream.

$sel:creationTime:LogStream', logStream_creationTime - The creation time of the stream, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

$sel:arn:LogStream', logStream_arn - The Amazon Resource Name (ARN) of the log stream.

$sel:storedBytes:LogStream', logStream_storedBytes - The number of bytes stored.

Important: On June 17, 2019, this parameter was deprecated for log streams, and is always reported as zero. This change applies only to log streams. The storedBytes parameter for log groups is not affected.

$sel:uploadSequenceToken:LogStream', logStream_uploadSequenceToken - The sequence token.

$sel:lastEventTimestamp:LogStream', logStream_lastEventTimestamp - The time of the most recent log event in the log stream in CloudWatch Logs. This number is expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC. The lastEventTime value updates on an eventual consistency basis. It typically updates in less than an hour from ingestion, but in rare situations might take longer.

$sel:firstEventTimestamp:LogStream', logStream_firstEventTimestamp - The time of the first event, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

$sel:lastIngestionTime:LogStream', logStream_lastIngestionTime - The ingestion time, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

logStream_logStreamName :: Lens' LogStream (Maybe Text)

The name of the log stream.

logStream_creationTime :: Lens' LogStream (Maybe Natural)

The creation time of the stream, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

logStream_arn :: Lens' LogStream (Maybe Text)

The Amazon Resource Name (ARN) of the log stream.

logStream_storedBytes :: Lens' LogStream (Maybe Natural)

The number of bytes stored.

Important: On June 17, 2019, this parameter was deprecated for log streams, and is always reported as zero. This change applies only to log streams. The storedBytes parameter for log groups is not affected.

logStream_uploadSequenceToken :: Lens' LogStream (Maybe Text)

The sequence token.

logStream_lastEventTimestamp :: Lens' LogStream (Maybe Natural)

The time of the most recent log event in the log stream in CloudWatch Logs. This number is expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC. The lastEventTime value updates on an eventual consistency basis. It typically updates in less than an hour from ingestion, but in rare situations might take longer.

logStream_firstEventTimestamp :: Lens' LogStream (Maybe Natural)

The time of the first event, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

logStream_lastIngestionTime :: Lens' LogStream (Maybe Natural)

The ingestion time, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.