Network.AWS.CloudWatchLogs.Types.QueryInfo

Description

Documentation

data QueryInfo

Information about one CloudWatch Logs Insights query that matches the request in a DescribeQueries operation.

See: newQueryInfo smart constructor.

Constructors

QueryInfo'

Fields

queryString :: Maybe Text

The query string used in this query.

status :: Maybe QueryStatus

The status of this query. Possible values are Cancelled, Complete, Failed, Running, Scheduled, and Unknown.

queryId :: Maybe Text

The unique ID number of this query.

logGroupName :: Maybe Text

The name of the log group scanned by this query.

createTime :: Maybe Natural

The date and time that this query was created.

newQueryInfo :: QueryInfo

Create a value of QueryInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:queryString:QueryInfo', queryInfo_queryString - The query string used in this query.

$sel:status:QueryInfo', queryInfo_status - The status of this query. Possible values are Cancelled, Complete, Failed, Running, Scheduled, and Unknown.

$sel:queryId:QueryInfo', queryInfo_queryId - The unique ID number of this query.

$sel:logGroupName:QueryInfo', queryInfo_logGroupName - The name of the log group scanned by this query.

$sel:createTime:QueryInfo', queryInfo_createTime - The date and time that this query was created.

queryInfo_queryString :: Lens' QueryInfo (Maybe Text)

The query string used in this query.

queryInfo_status :: Lens' QueryInfo (Maybe QueryStatus)

The status of this query. Possible values are Cancelled, Complete, Failed, Running, Scheduled, and Unknown.

queryInfo_queryId :: Lens' QueryInfo (Maybe Text)

The unique ID number of this query.

queryInfo_logGroupName :: Lens' QueryInfo (Maybe Text)

The name of the log group scanned by this query.

queryInfo_createTime :: Lens' QueryInfo (Maybe Natural)

The date and time that this query was created.