Network.AWS.CloudWatchLogs.Types.ResourcePolicy

Description

Documentation

data ResourcePolicy

A policy enabling one or more entities to put logs to a log group in this account.

See: newResourcePolicy smart constructor.

Constructors

ResourcePolicy'

Fields

policyDocument :: Maybe Text

The details of the policy.

policyName :: Maybe Text

The name of the resource policy.

lastUpdatedTime :: Maybe Natural

Timestamp showing when this policy was last updated, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

newResourcePolicy :: ResourcePolicy

Create a value of ResourcePolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policyDocument:ResourcePolicy', resourcePolicy_policyDocument - The details of the policy.

$sel:policyName:ResourcePolicy', resourcePolicy_policyName - The name of the resource policy.

$sel:lastUpdatedTime:ResourcePolicy', resourcePolicy_lastUpdatedTime - Timestamp showing when this policy was last updated, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

resourcePolicy_policyDocument :: Lens' ResourcePolicy (Maybe Text)

The details of the policy.

resourcePolicy_policyName :: Lens' ResourcePolicy (Maybe Text)

The name of the resource policy.

resourcePolicy_lastUpdatedTime :: Lens' ResourcePolicy (Maybe Natural)

Timestamp showing when this policy was last updated, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.