Network.AWS.CodeArtifact.AssociateExternalConnection

Description

Adds an existing external connection to a repository. One external connection is allowed per repository.

A repository can have one or more upstream repositories, or an external connection.

Creating a Request

data AssociateExternalConnection

See: newAssociateExternalConnection smart constructor.

Constructors

AssociateExternalConnection'

Fields

domainOwner :: Maybe Text

The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

domain :: Text

The name of the domain that contains the repository.

repository :: Text

The name of the repository to which the external connection is added.

externalConnection :: Text

The name of the external connection to add to the repository. The following values are supported:

public:npmjs - for the npm public repository.

public:pypi - for the Python Package Index.

public:maven-central - for Maven Central.

public:maven-googleandroid - for the Google Android repository.

public:maven-gradleplugins - for the Gradle plugins repository.

public:maven-commonsware - for the CommonsWare Android repository.

newAssociateExternalConnection

Arguments

:: Text
$sel:domain:AssociateExternalConnection'

-> Text
$sel:repository:AssociateExternalConnection'

-> Text
$sel:externalConnection:AssociateExternalConnection'

-> AssociateExternalConnection

Create a value of AssociateExternalConnection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainOwner:AssociateExternalConnection', associateExternalConnection_domainOwner - The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

$sel:domain:AssociateExternalConnection', associateExternalConnection_domain - The name of the domain that contains the repository.

$sel:repository:AssociateExternalConnection', associateExternalConnection_repository - The name of the repository to which the external connection is added.

$sel:externalConnection:AssociateExternalConnection', associateExternalConnection_externalConnection - The name of the external connection to add to the repository. The following values are supported:

public:npmjs - for the npm public repository.

public:pypi - for the Python Package Index.

public:maven-central - for Maven Central.

public:maven-googleandroid - for the Google Android repository.

public:maven-gradleplugins - for the Gradle plugins repository.

public:maven-commonsware - for the CommonsWare Android repository.

Request Lenses

associateExternalConnection_domainOwner :: Lens' AssociateExternalConnection (Maybe Text)

The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

associateExternalConnection_domain :: Lens' AssociateExternalConnection Text

The name of the domain that contains the repository.

associateExternalConnection_repository :: Lens' AssociateExternalConnection Text

The name of the repository to which the external connection is added.

associateExternalConnection_externalConnection :: Lens' AssociateExternalConnection Text

The name of the external connection to add to the repository. The following values are supported:

public:npmjs - for the npm public repository.

public:pypi - for the Python Package Index.

public:maven-central - for Maven Central.

public:maven-googleandroid - for the Google Android repository.

public:maven-gradleplugins - for the Gradle plugins repository.

public:maven-commonsware - for the CommonsWare Android repository.

Destructuring the Response

data AssociateExternalConnectionResponse

See: newAssociateExternalConnectionResponse smart constructor.

Constructors

AssociateExternalConnectionResponse'

Fields

repository :: Maybe RepositoryDescription

Information about the connected repository after processing the request.

httpStatus :: Int

The response's http status code.

newAssociateExternalConnectionResponse

Arguments

:: Int
$sel:httpStatus:AssociateExternalConnectionResponse'

-> AssociateExternalConnectionResponse

Create a value of AssociateExternalConnectionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:repository:AssociateExternalConnection', associateExternalConnectionResponse_repository - Information about the connected repository after processing the request.

$sel:httpStatus:AssociateExternalConnectionResponse', associateExternalConnectionResponse_httpStatus - The response's http status code.

Response Lenses

associateExternalConnectionResponse_repository :: Lens' AssociateExternalConnectionResponse (Maybe RepositoryDescription)

Information about the connected repository after processing the request.

associateExternalConnectionResponse_httpStatus :: Lens' AssociateExternalConnectionResponse Int

The response's http status code.