Network.AWS.CodeArtifact.CreateRepository

Description

Creates a repository.

Creating a Request

data CreateRepository

See: newCreateRepository smart constructor.

Constructors

CreateRepository'

Fields

upstreams :: Maybe [UpstreamRepository]

A list of upstream repositories to associate with the repository. The order of the upstream repositories in the list determines their priority order when AWS CodeArtifact looks for a requested package version. For more information, see Working with upstream repositories.

domainOwner :: Maybe Text

The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

description :: Maybe Text

A description of the created repository.

tags :: Maybe [Tag]

One or more tag key-value pairs for the repository.

domain :: Text

The name of the domain that contains the created repository.

repository :: Text

The name of the repository to create.

newCreateRepository

Arguments

:: Text
$sel:domain:CreateRepository'

-> Text
$sel:repository:CreateRepository'

-> CreateRepository

Create a value of CreateRepository with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:upstreams:CreateRepository', createRepository_upstreams - A list of upstream repositories to associate with the repository. The order of the upstream repositories in the list determines their priority order when AWS CodeArtifact looks for a requested package version. For more information, see Working with upstream repositories.

$sel:domainOwner:CreateRepository', createRepository_domainOwner - The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

$sel:description:CreateRepository', createRepository_description - A description of the created repository.

$sel:tags:CreateRepository', createRepository_tags - One or more tag key-value pairs for the repository.

$sel:domain:CreateRepository', createRepository_domain - The name of the domain that contains the created repository.

$sel:repository:CreateRepository', createRepository_repository - The name of the repository to create.

Request Lenses

createRepository_upstreams :: Lens' CreateRepository (Maybe [UpstreamRepository])

A list of upstream repositories to associate with the repository. The order of the upstream repositories in the list determines their priority order when AWS CodeArtifact looks for a requested package version. For more information, see Working with upstream repositories.

createRepository_domainOwner :: Lens' CreateRepository (Maybe Text)

The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

createRepository_description :: Lens' CreateRepository (Maybe Text)

A description of the created repository.

createRepository_tags :: Lens' CreateRepository (Maybe [Tag])

One or more tag key-value pairs for the repository.

createRepository_domain :: Lens' CreateRepository Text

The name of the domain that contains the created repository.

createRepository_repository :: Lens' CreateRepository Text

The name of the repository to create.

Destructuring the Response

data CreateRepositoryResponse

See: newCreateRepositoryResponse smart constructor.

Constructors

CreateRepositoryResponse'

Fields

repository :: Maybe RepositoryDescription

Information about the created repository after processing the request.

httpStatus :: Int

The response's http status code.

newCreateRepositoryResponse

Arguments

:: Int
$sel:httpStatus:CreateRepositoryResponse'

-> CreateRepositoryResponse

Create a value of CreateRepositoryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:repository:CreateRepository', createRepositoryResponse_repository - Information about the created repository after processing the request.

$sel:httpStatus:CreateRepositoryResponse', createRepositoryResponse_httpStatus - The response's http status code.

Response Lenses

createRepositoryResponse_repository :: Lens' CreateRepositoryResponse (Maybe RepositoryDescription)

Information about the created repository after processing the request.

createRepositoryResponse_httpStatus :: Lens' CreateRepositoryResponse Int

The response's http status code.