Network.AWS.CodeArtifact.DeleteRepositoryPermissionsPolicy

Description

Deletes the resource policy that is set on a repository. After a resource policy is deleted, the permissions allowed and denied by the deleted policy are removed. The effect of deleting a resource policy might not be immediate.

Use DeleteRepositoryPermissionsPolicy with caution. After a policy is deleted, AWS users, roles, and accounts lose permissions to perform the repository actions granted by the deleted policy.

Creating a Request

data DeleteRepositoryPermissionsPolicy

See: newDeleteRepositoryPermissionsPolicy smart constructor.

Constructors

DeleteRepositoryPermissionsPolicy'

Fields

policyRevision :: Maybe Text

The revision of the repository's resource policy to be deleted. This revision is used for optimistic locking, which prevents others from accidentally overwriting your changes to the repository's resource policy.

domainOwner :: Maybe Text

The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

domain :: Text

The name of the domain that contains the repository associated with the resource policy to be deleted.

repository :: Text

The name of the repository that is associated with the resource policy to be deleted

newDeleteRepositoryPermissionsPolicy

Arguments

:: Text
$sel:domain:DeleteRepositoryPermissionsPolicy'

-> Text
$sel:repository:DeleteRepositoryPermissionsPolicy'

-> DeleteRepositoryPermissionsPolicy

Create a value of DeleteRepositoryPermissionsPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policyRevision:DeleteRepositoryPermissionsPolicy', deleteRepositoryPermissionsPolicy_policyRevision - The revision of the repository's resource policy to be deleted. This revision is used for optimistic locking, which prevents others from accidentally overwriting your changes to the repository's resource policy.

$sel:domainOwner:DeleteRepositoryPermissionsPolicy', deleteRepositoryPermissionsPolicy_domainOwner - The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

$sel:domain:DeleteRepositoryPermissionsPolicy', deleteRepositoryPermissionsPolicy_domain - The name of the domain that contains the repository associated with the resource policy to be deleted.

$sel:repository:DeleteRepositoryPermissionsPolicy', deleteRepositoryPermissionsPolicy_repository - The name of the repository that is associated with the resource policy to be deleted

Request Lenses

deleteRepositoryPermissionsPolicy_policyRevision :: Lens' DeleteRepositoryPermissionsPolicy (Maybe Text)

The revision of the repository's resource policy to be deleted. This revision is used for optimistic locking, which prevents others from accidentally overwriting your changes to the repository's resource policy.

deleteRepositoryPermissionsPolicy_domainOwner :: Lens' DeleteRepositoryPermissionsPolicy (Maybe Text)

The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

deleteRepositoryPermissionsPolicy_domain :: Lens' DeleteRepositoryPermissionsPolicy Text

The name of the domain that contains the repository associated with the resource policy to be deleted.

deleteRepositoryPermissionsPolicy_repository :: Lens' DeleteRepositoryPermissionsPolicy Text

The name of the repository that is associated with the resource policy to be deleted

Destructuring the Response

data DeleteRepositoryPermissionsPolicyResponse

See: newDeleteRepositoryPermissionsPolicyResponse smart constructor.

Constructors

DeleteRepositoryPermissionsPolicyResponse'

Fields

policy :: Maybe ResourcePolicy

Information about the deleted policy after processing the request.

httpStatus :: Int

The response's http status code.

newDeleteRepositoryPermissionsPolicyResponse

Arguments

:: Int
$sel:httpStatus:DeleteRepositoryPermissionsPolicyResponse'

-> DeleteRepositoryPermissionsPolicyResponse

Create a value of DeleteRepositoryPermissionsPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policy:DeleteRepositoryPermissionsPolicyResponse', deleteRepositoryPermissionsPolicyResponse_policy - Information about the deleted policy after processing the request.

$sel:httpStatus:DeleteRepositoryPermissionsPolicyResponse', deleteRepositoryPermissionsPolicyResponse_httpStatus - The response's http status code.

Response Lenses

deleteRepositoryPermissionsPolicyResponse_policy :: Lens' DeleteRepositoryPermissionsPolicyResponse (Maybe ResourcePolicy)

Information about the deleted policy after processing the request.

deleteRepositoryPermissionsPolicyResponse_httpStatus :: Lens' DeleteRepositoryPermissionsPolicyResponse Int

The response's http status code.