Network.AWS.CodeArtifact.GetPackageVersionAsset

Description

Returns an asset (or file) that is in a package. For example, for a Maven package version, use GetPackageVersionAsset to download a JAR file, a POM file, or any other assets in the package version.

Creating a Request

data GetPackageVersionAsset

See: newGetPackageVersionAsset smart constructor.

Constructors

GetPackageVersionAsset'

Fields

namespace :: Maybe Text

The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

domainOwner :: Maybe Text

The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

packageVersionRevision :: Maybe Text

The name of the package version revision that contains the requested asset.

domain :: Text

The name of the domain that contains the repository that contains the package version with the requested asset.

repository :: Text

The repository that contains the package version with the requested asset.

format :: PackageFormat

A format that specifies the type of the package version with the requested asset file. The valid values are:

npm

pypi

maven

package :: Text

The name of the package that contains the requested asset.

packageVersion :: Text

A string that contains the package version (for example, 3.5.2).

asset :: Text

The name of the requested asset.

newGetPackageVersionAsset

Arguments

:: Text
$sel:domain:GetPackageVersionAsset'

-> Text
$sel:repository:GetPackageVersionAsset'

-> PackageFormat
$sel:format:GetPackageVersionAsset'

-> Text
$sel:package:GetPackageVersionAsset'

-> Text
$sel:packageVersion:GetPackageVersionAsset'

-> Text
$sel:asset:GetPackageVersionAsset'

-> GetPackageVersionAsset

Create a value of GetPackageVersionAsset with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:namespace:GetPackageVersionAsset', getPackageVersionAsset_namespace - The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

$sel:domainOwner:GetPackageVersionAsset', getPackageVersionAsset_domainOwner - The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

$sel:packageVersionRevision:GetPackageVersionAsset', getPackageVersionAsset_packageVersionRevision - The name of the package version revision that contains the requested asset.

$sel:domain:GetPackageVersionAsset', getPackageVersionAsset_domain - The name of the domain that contains the repository that contains the package version with the requested asset.

$sel:repository:GetPackageVersionAsset', getPackageVersionAsset_repository - The repository that contains the package version with the requested asset.

$sel:format:GetPackageVersionAsset', getPackageVersionAsset_format - A format that specifies the type of the package version with the requested asset file. The valid values are:

npm

pypi

maven

$sel:package:GetPackageVersionAsset', getPackageVersionAsset_package - The name of the package that contains the requested asset.

$sel:packageVersion:GetPackageVersionAsset', getPackageVersionAsset_packageVersion - A string that contains the package version (for example, 3.5.2).

$sel:asset:GetPackageVersionAsset', getPackageVersionAsset_asset - The name of the requested asset.

Request Lenses

getPackageVersionAsset_namespace :: Lens' GetPackageVersionAsset (Maybe Text)

The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

getPackageVersionAsset_domainOwner :: Lens' GetPackageVersionAsset (Maybe Text)

The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

getPackageVersionAsset_packageVersionRevision :: Lens' GetPackageVersionAsset (Maybe Text)

The name of the package version revision that contains the requested asset.

getPackageVersionAsset_domain :: Lens' GetPackageVersionAsset Text

The name of the domain that contains the repository that contains the package version with the requested asset.

getPackageVersionAsset_repository :: Lens' GetPackageVersionAsset Text

The repository that contains the package version with the requested asset.

getPackageVersionAsset_format :: Lens' GetPackageVersionAsset PackageFormat

A format that specifies the type of the package version with the requested asset file. The valid values are:

npm

pypi

maven

getPackageVersionAsset_package :: Lens' GetPackageVersionAsset Text

The name of the package that contains the requested asset.

getPackageVersionAsset_packageVersion :: Lens' GetPackageVersionAsset Text

A string that contains the package version (for example, 3.5.2).

getPackageVersionAsset_asset :: Lens' GetPackageVersionAsset Text

The name of the requested asset.

Destructuring the Response

data GetPackageVersionAssetResponse

See: newGetPackageVersionAssetResponse smart constructor.

Constructors

GetPackageVersionAssetResponse'

Fields

packageVersion :: Maybe Text

A string that contains the package version (for example, 3.5.2).

assetName :: Maybe Text

The name of the asset that is downloaded.

packageVersionRevision :: Maybe Text

The name of the package version revision that contains the downloaded asset.

httpStatus :: Int

The response's http status code.

asset :: ResponseBody

The binary file, or asset, that is downloaded.

newGetPackageVersionAssetResponse

Arguments

:: Int
$sel:httpStatus:GetPackageVersionAssetResponse'

-> ResponseBody
$sel:asset:GetPackageVersionAsset'

-> GetPackageVersionAssetResponse

Create a value of GetPackageVersionAssetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:packageVersion:GetPackageVersionAsset', getPackageVersionAssetResponse_packageVersion - A string that contains the package version (for example, 3.5.2).

$sel:assetName:GetPackageVersionAssetResponse', getPackageVersionAssetResponse_assetName - The name of the asset that is downloaded.

$sel:packageVersionRevision:GetPackageVersionAsset', getPackageVersionAssetResponse_packageVersionRevision - The name of the package version revision that contains the downloaded asset.

$sel:httpStatus:GetPackageVersionAssetResponse', getPackageVersionAssetResponse_httpStatus - The response's http status code.

$sel:asset:GetPackageVersionAsset', getPackageVersionAssetResponse_asset - The binary file, or asset, that is downloaded.

Response Lenses

getPackageVersionAssetResponse_packageVersion :: Lens' GetPackageVersionAssetResponse (Maybe Text)

A string that contains the package version (for example, 3.5.2).

getPackageVersionAssetResponse_assetName :: Lens' GetPackageVersionAssetResponse (Maybe Text)

The name of the asset that is downloaded.

getPackageVersionAssetResponse_packageVersionRevision :: Lens' GetPackageVersionAssetResponse (Maybe Text)

The name of the package version revision that contains the downloaded asset.

getPackageVersionAssetResponse_httpStatus :: Lens' GetPackageVersionAssetResponse Int

The response's http status code.

getPackageVersionAssetResponse_asset :: Lens' GetPackageVersionAssetResponse ResponseBody

The binary file, or asset, that is downloaded.