Network.AWS.CodeArtifact.GetPackageVersionReadme

Description

Gets the readme file or descriptive text for a package version. For packages that do not contain a readme file, CodeArtifact extracts a description from a metadata file. For example, from the <description> element in the pom.xml file of a Maven package.

The returned text might contain formatting. For example, it might contain formatting for Markdown or reStructuredText.

Creating a Request

data GetPackageVersionReadme

See: newGetPackageVersionReadme smart constructor.

Constructors

GetPackageVersionReadme'

Fields

namespace :: Maybe Text

The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

domainOwner :: Maybe Text

The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

domain :: Text

The name of the domain that contains the repository that contains the package version with the requested readme file.

repository :: Text

The repository that contains the package with the requested readme file.

format :: PackageFormat

A format that specifies the type of the package version with the requested readme file. The valid values are:

npm

pypi

maven

package :: Text

The name of the package version that contains the requested readme file.

packageVersion :: Text

A string that contains the package version (for example, 3.5.2).

newGetPackageVersionReadme

Arguments

:: Text
$sel:domain:GetPackageVersionReadme'

-> Text
$sel:repository:GetPackageVersionReadme'

-> PackageFormat
$sel:format:GetPackageVersionReadme'

-> Text
$sel:package:GetPackageVersionReadme'

-> Text
$sel:packageVersion:GetPackageVersionReadme'

-> GetPackageVersionReadme

Create a value of GetPackageVersionReadme with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:namespace:GetPackageVersionReadme', getPackageVersionReadme_namespace - The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

$sel:domainOwner:GetPackageVersionReadme', getPackageVersionReadme_domainOwner - The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

$sel:domain:GetPackageVersionReadme', getPackageVersionReadme_domain - The name of the domain that contains the repository that contains the package version with the requested readme file.

$sel:repository:GetPackageVersionReadme', getPackageVersionReadme_repository - The repository that contains the package with the requested readme file.

$sel:format:GetPackageVersionReadme', getPackageVersionReadme_format - A format that specifies the type of the package version with the requested readme file. The valid values are:

npm

pypi

maven

$sel:package:GetPackageVersionReadme', getPackageVersionReadme_package - The name of the package version that contains the requested readme file.

$sel:packageVersion:GetPackageVersionReadme', getPackageVersionReadme_packageVersion - A string that contains the package version (for example, 3.5.2).

Request Lenses

getPackageVersionReadme_namespace :: Lens' GetPackageVersionReadme (Maybe Text)

The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

getPackageVersionReadme_domainOwner :: Lens' GetPackageVersionReadme (Maybe Text)

The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

getPackageVersionReadme_domain :: Lens' GetPackageVersionReadme Text

The name of the domain that contains the repository that contains the package version with the requested readme file.

getPackageVersionReadme_repository :: Lens' GetPackageVersionReadme Text

The repository that contains the package with the requested readme file.

getPackageVersionReadme_format :: Lens' GetPackageVersionReadme PackageFormat

A format that specifies the type of the package version with the requested readme file. The valid values are:

npm

pypi

maven

getPackageVersionReadme_package :: Lens' GetPackageVersionReadme Text

The name of the package version that contains the requested readme file.

getPackageVersionReadme_packageVersion :: Lens' GetPackageVersionReadme Text

A string that contains the package version (for example, 3.5.2).

Destructuring the Response

data GetPackageVersionReadmeResponse

See: newGetPackageVersionReadmeResponse smart constructor.

Constructors

GetPackageVersionReadmeResponse'

Fields

readme :: Maybe Text

The text of the returned readme file.

format :: Maybe PackageFormat

The format of the package with the requested readme file. Valid format types are:

npm

pypi

maven

namespace :: Maybe Text

The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

versionRevision :: Maybe Text

The current revision associated with the package version.

version :: Maybe Text

The version of the package with the requested readme file.

package :: Maybe Text

The name of the package that contains the returned readme file.

httpStatus :: Int

The response's http status code.

newGetPackageVersionReadmeResponse

Arguments

:: Int
$sel:httpStatus:GetPackageVersionReadmeResponse'

-> GetPackageVersionReadmeResponse

Create a value of GetPackageVersionReadmeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:readme:GetPackageVersionReadmeResponse', getPackageVersionReadmeResponse_readme - The text of the returned readme file.

$sel:format:GetPackageVersionReadme', getPackageVersionReadmeResponse_format - The format of the package with the requested readme file. Valid format types are:

npm

pypi

maven

$sel:namespace:GetPackageVersionReadme', getPackageVersionReadmeResponse_namespace - The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

$sel:versionRevision:GetPackageVersionReadmeResponse', getPackageVersionReadmeResponse_versionRevision - The current revision associated with the package version.

$sel:version:GetPackageVersionReadmeResponse', getPackageVersionReadmeResponse_version - The version of the package with the requested readme file.

$sel:package:GetPackageVersionReadme', getPackageVersionReadmeResponse_package - The name of the package that contains the returned readme file.

$sel:httpStatus:GetPackageVersionReadmeResponse', getPackageVersionReadmeResponse_httpStatus - The response's http status code.

Response Lenses

getPackageVersionReadmeResponse_readme :: Lens' GetPackageVersionReadmeResponse (Maybe Text)

The text of the returned readme file.

getPackageVersionReadmeResponse_format :: Lens' GetPackageVersionReadmeResponse (Maybe PackageFormat)

The format of the package with the requested readme file. Valid format types are:

npm

pypi

maven

getPackageVersionReadmeResponse_namespace :: Lens' GetPackageVersionReadmeResponse (Maybe Text)

The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

getPackageVersionReadmeResponse_versionRevision :: Lens' GetPackageVersionReadmeResponse (Maybe Text)

The current revision associated with the package version.

getPackageVersionReadmeResponse_version :: Lens' GetPackageVersionReadmeResponse (Maybe Text)

The version of the package with the requested readme file.

getPackageVersionReadmeResponse_package :: Lens' GetPackageVersionReadmeResponse (Maybe Text)

The name of the package that contains the returned readme file.

getPackageVersionReadmeResponse_httpStatus :: Lens' GetPackageVersionReadmeResponse Int

The response's http status code.