Network.AWS.CodeArtifact.ListPackageVersionDependencies

Description

Returns the direct dependencies for a package version. The dependencies are returned as PackageDependency objects. CodeArtifact extracts the dependencies for a package version from the metadata file for the package format (for example, the package.json file for npm packages and the pom.xml file for Maven). Any package version dependencies that are not listed in the configuration file are not returned.

Creating a Request

data ListPackageVersionDependencies

See: newListPackageVersionDependencies smart constructor.

Constructors

ListPackageVersionDependencies'

Fields

namespace :: Maybe Text

The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

domainOwner :: Maybe Text

The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

nextToken :: Maybe Text

The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

domain :: Text

The name of the domain that contains the repository that contains the requested package version dependencies.

repository :: Text

The name of the repository that contains the requested package version.

format :: PackageFormat

The format of the package with the requested dependencies. The valid package types are:

npm: A Node Package Manager (npm) package.

pypi: A Python Package Index (PyPI) package.

maven: A Maven package that contains compiled code in a distributable format, such as a JAR file.

package :: Text

The name of the package versions' package.

packageVersion :: Text

A string that contains the package version (for example, 3.5.2).

newListPackageVersionDependencies

Arguments

:: Text
$sel:domain:ListPackageVersionDependencies'

-> Text
$sel:repository:ListPackageVersionDependencies'

-> PackageFormat
$sel:format:ListPackageVersionDependencies'

-> Text
$sel:package:ListPackageVersionDependencies'

-> Text
$sel:packageVersion:ListPackageVersionDependencies'

-> ListPackageVersionDependencies

Create a value of ListPackageVersionDependencies with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:namespace:ListPackageVersionDependencies', listPackageVersionDependencies_namespace - The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

$sel:domainOwner:ListPackageVersionDependencies', listPackageVersionDependencies_domainOwner - The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

$sel:nextToken:ListPackageVersionDependencies', listPackageVersionDependencies_nextToken - The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

$sel:domain:ListPackageVersionDependencies', listPackageVersionDependencies_domain - The name of the domain that contains the repository that contains the requested package version dependencies.

$sel:repository:ListPackageVersionDependencies', listPackageVersionDependencies_repository - The name of the repository that contains the requested package version.

$sel:format:ListPackageVersionDependencies', listPackageVersionDependencies_format - The format of the package with the requested dependencies. The valid package types are:

npm: A Node Package Manager (npm) package.

pypi: A Python Package Index (PyPI) package.

maven: A Maven package that contains compiled code in a distributable format, such as a JAR file.

$sel:package:ListPackageVersionDependencies', listPackageVersionDependencies_package - The name of the package versions' package.

$sel:packageVersion:ListPackageVersionDependencies', listPackageVersionDependencies_packageVersion - A string that contains the package version (for example, 3.5.2).

Request Lenses

listPackageVersionDependencies_namespace :: Lens' ListPackageVersionDependencies (Maybe Text)

The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

listPackageVersionDependencies_domainOwner :: Lens' ListPackageVersionDependencies (Maybe Text)

The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

listPackageVersionDependencies_nextToken :: Lens' ListPackageVersionDependencies (Maybe Text)

The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

listPackageVersionDependencies_domain :: Lens' ListPackageVersionDependencies Text

The name of the domain that contains the repository that contains the requested package version dependencies.

listPackageVersionDependencies_repository :: Lens' ListPackageVersionDependencies Text

The name of the repository that contains the requested package version.

listPackageVersionDependencies_format :: Lens' ListPackageVersionDependencies PackageFormat

The format of the package with the requested dependencies. The valid package types are:

npm: A Node Package Manager (npm) package.

pypi: A Python Package Index (PyPI) package.

maven: A Maven package that contains compiled code in a distributable format, such as a JAR file.

listPackageVersionDependencies_package :: Lens' ListPackageVersionDependencies Text

The name of the package versions' package.

listPackageVersionDependencies_packageVersion :: Lens' ListPackageVersionDependencies Text

A string that contains the package version (for example, 3.5.2).

Destructuring the Response

data ListPackageVersionDependenciesResponse

See: newListPackageVersionDependenciesResponse smart constructor.

Constructors

ListPackageVersionDependenciesResponse'

Fields

format :: Maybe PackageFormat

A format that specifies the type of the package that contains the returned dependencies. The valid values are:

npm

pypi

maven

namespace :: Maybe Text

The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

versionRevision :: Maybe Text

The current revision associated with the package version.

dependencies :: Maybe [PackageDependency]

The returned list of PackageDependency objects.

nextToken :: Maybe Text

The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

version :: Maybe Text

The version of the package that is specified in the request.

package :: Maybe Text

The name of the package that contains the returned package versions dependencies.

httpStatus :: Int

The response's http status code.

newListPackageVersionDependenciesResponse

Arguments

:: Int
$sel:httpStatus:ListPackageVersionDependenciesResponse'

-> ListPackageVersionDependenciesResponse

Create a value of ListPackageVersionDependenciesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:format:ListPackageVersionDependencies', listPackageVersionDependenciesResponse_format - A format that specifies the type of the package that contains the returned dependencies. The valid values are:

npm

pypi

maven

$sel:namespace:ListPackageVersionDependencies', listPackageVersionDependenciesResponse_namespace - The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

$sel:versionRevision:ListPackageVersionDependenciesResponse', listPackageVersionDependenciesResponse_versionRevision - The current revision associated with the package version.

$sel:dependencies:ListPackageVersionDependenciesResponse', listPackageVersionDependenciesResponse_dependencies - The returned list of PackageDependency objects.

$sel:nextToken:ListPackageVersionDependencies', listPackageVersionDependenciesResponse_nextToken - The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

$sel:version:ListPackageVersionDependenciesResponse', listPackageVersionDependenciesResponse_version - The version of the package that is specified in the request.

$sel:package:ListPackageVersionDependencies', listPackageVersionDependenciesResponse_package - The name of the package that contains the returned package versions dependencies.

$sel:httpStatus:ListPackageVersionDependenciesResponse', listPackageVersionDependenciesResponse_httpStatus - The response's http status code.

Response Lenses

listPackageVersionDependenciesResponse_format :: Lens' ListPackageVersionDependenciesResponse (Maybe PackageFormat)

A format that specifies the type of the package that contains the returned dependencies. The valid values are:

npm

pypi

maven

listPackageVersionDependenciesResponse_namespace :: Lens' ListPackageVersionDependenciesResponse (Maybe Text)

The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

listPackageVersionDependenciesResponse_versionRevision :: Lens' ListPackageVersionDependenciesResponse (Maybe Text)

The current revision associated with the package version.

listPackageVersionDependenciesResponse_dependencies :: Lens' ListPackageVersionDependenciesResponse (Maybe [PackageDependency])

The returned list of PackageDependency objects.

listPackageVersionDependenciesResponse_nextToken :: Lens' ListPackageVersionDependenciesResponse (Maybe Text)

The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

listPackageVersionDependenciesResponse_version :: Lens' ListPackageVersionDependenciesResponse (Maybe Text)

The version of the package that is specified in the request.

listPackageVersionDependenciesResponse_package :: Lens' ListPackageVersionDependenciesResponse (Maybe Text)

The name of the package that contains the returned package versions dependencies.

listPackageVersionDependenciesResponse_httpStatus :: Lens' ListPackageVersionDependenciesResponse Int

The response's http status code.