Network.AWS.CodeArtifact.ListPackageVersions

Description

Returns a list of PackageVersionSummary objects for package versions in a repository that match the request parameters.

This operation returns paginated results.

Creating a Request

data ListPackageVersions

See: newListPackageVersions smart constructor.

Constructors

ListPackageVersions'

Fields

status :: Maybe PackageVersionStatus

A string that specifies the status of the package versions to include in the returned list. It can be one of the following:

Published

Unfinished

Unlisted

Archived

Disposed

namespace :: Maybe Text

The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

domainOwner :: Maybe Text

The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

nextToken :: Maybe Text

The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

maxResults :: Maybe Natural

The maximum number of results to return per page.

sortBy :: Maybe PackageVersionSortType

How to sort the returned list of package versions.

domain :: Text

The name of the domain that contains the repository that contains the returned package versions.

repository :: Text

The name of the repository that contains the package.

format :: PackageFormat

The format of the returned packages. The valid package types are:

npm: A Node Package Manager (npm) package.

pypi: A Python Package Index (PyPI) package.

maven: A Maven package that contains compiled code in a distributable format, such as a JAR file.

package :: Text

The name of the package for which you want to return a list of package versions.

newListPackageVersions

Arguments

:: Text
$sel:domain:ListPackageVersions'

-> Text
$sel:repository:ListPackageVersions'

-> PackageFormat
$sel:format:ListPackageVersions'

-> Text
$sel:package:ListPackageVersions'

-> ListPackageVersions

Create a value of ListPackageVersions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ListPackageVersions', listPackageVersions_status - A string that specifies the status of the package versions to include in the returned list. It can be one of the following:

Published

Unfinished

Unlisted

Archived

Disposed

$sel:namespace:ListPackageVersions', listPackageVersions_namespace - The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

$sel:domainOwner:ListPackageVersions', listPackageVersions_domainOwner - The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

$sel:nextToken:ListPackageVersions', listPackageVersions_nextToken - The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

$sel:maxResults:ListPackageVersions', listPackageVersions_maxResults - The maximum number of results to return per page.

$sel:sortBy:ListPackageVersions', listPackageVersions_sortBy - How to sort the returned list of package versions.

$sel:domain:ListPackageVersions', listPackageVersions_domain - The name of the domain that contains the repository that contains the returned package versions.

$sel:repository:ListPackageVersions', listPackageVersions_repository - The name of the repository that contains the package.

$sel:format:ListPackageVersions', listPackageVersions_format - The format of the returned packages. The valid package types are:

npm: A Node Package Manager (npm) package.

pypi: A Python Package Index (PyPI) package.

maven: A Maven package that contains compiled code in a distributable format, such as a JAR file.

$sel:package:ListPackageVersions', listPackageVersions_package - The name of the package for which you want to return a list of package versions.

Request Lenses

listPackageVersions_status :: Lens' ListPackageVersions (Maybe PackageVersionStatus)

A string that specifies the status of the package versions to include in the returned list. It can be one of the following:

Published

Unfinished

Unlisted

Archived

Disposed

listPackageVersions_namespace :: Lens' ListPackageVersions (Maybe Text)

The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

listPackageVersions_domainOwner :: Lens' ListPackageVersions (Maybe Text)

The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

listPackageVersions_nextToken :: Lens' ListPackageVersions (Maybe Text)

The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

listPackageVersions_maxResults :: Lens' ListPackageVersions (Maybe Natural)

The maximum number of results to return per page.

listPackageVersions_sortBy :: Lens' ListPackageVersions (Maybe PackageVersionSortType)

How to sort the returned list of package versions.

listPackageVersions_domain :: Lens' ListPackageVersions Text

The name of the domain that contains the repository that contains the returned package versions.

listPackageVersions_repository :: Lens' ListPackageVersions Text

The name of the repository that contains the package.

listPackageVersions_format :: Lens' ListPackageVersions PackageFormat

The format of the returned packages. The valid package types are:

npm: A Node Package Manager (npm) package.

pypi: A Python Package Index (PyPI) package.

maven: A Maven package that contains compiled code in a distributable format, such as a JAR file.

listPackageVersions_package :: Lens' ListPackageVersions Text

The name of the package for which you want to return a list of package versions.

Destructuring the Response

data ListPackageVersionsResponse

See: newListPackageVersionsResponse smart constructor.

Constructors

ListPackageVersionsResponse'

Fields

format :: Maybe PackageFormat

A format of the package. Valid package format values are:

npm

pypi

maven

namespace :: Maybe Text

The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

versions :: Maybe [PackageVersionSummary]

The returned list of PackageVersionSummary objects.

nextToken :: Maybe Text

If there are additional results, this is the token for the next set of results.

package :: Maybe Text

The name of the package.

defaultDisplayVersion :: Maybe Text

The default package version to display. This depends on the package format:

For Maven and PyPI packages, it's the most recently published package version.

For npm packages, it's the version referenced by the latest tag. If the latest tag is not set, it's the most recently published package version.

httpStatus :: Int

The response's http status code.

newListPackageVersionsResponse

Arguments

:: Int
$sel:httpStatus:ListPackageVersionsResponse'

-> ListPackageVersionsResponse

Create a value of ListPackageVersionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:format:ListPackageVersions', listPackageVersionsResponse_format - A format of the package. Valid package format values are:

npm

pypi

maven

$sel:namespace:ListPackageVersions', listPackageVersionsResponse_namespace - The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

$sel:versions:ListPackageVersionsResponse', listPackageVersionsResponse_versions - The returned list of PackageVersionSummary objects.

$sel:nextToken:ListPackageVersions', listPackageVersionsResponse_nextToken - If there are additional results, this is the token for the next set of results.

$sel:package:ListPackageVersions', listPackageVersionsResponse_package - The name of the package.

$sel:defaultDisplayVersion:ListPackageVersionsResponse', listPackageVersionsResponse_defaultDisplayVersion - The default package version to display. This depends on the package format:

For Maven and PyPI packages, it's the most recently published package version.

For npm packages, it's the version referenced by the latest tag. If the latest tag is not set, it's the most recently published package version.

$sel:httpStatus:ListPackageVersionsResponse', listPackageVersionsResponse_httpStatus - The response's http status code.

Response Lenses

listPackageVersionsResponse_format :: Lens' ListPackageVersionsResponse (Maybe PackageFormat)

A format of the package. Valid package format values are:

npm

pypi

maven

listPackageVersionsResponse_namespace :: Lens' ListPackageVersionsResponse (Maybe Text)

The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

listPackageVersionsResponse_versions :: Lens' ListPackageVersionsResponse (Maybe [PackageVersionSummary])

The returned list of PackageVersionSummary objects.

listPackageVersionsResponse_nextToken :: Lens' ListPackageVersionsResponse (Maybe Text)

If there are additional results, this is the token for the next set of results.

listPackageVersionsResponse_package :: Lens' ListPackageVersionsResponse (Maybe Text)

The name of the package.

listPackageVersionsResponse_defaultDisplayVersion :: Lens' ListPackageVersionsResponse (Maybe Text)

The default package version to display. This depends on the package format:

For Maven and PyPI packages, it's the most recently published package version.

For npm packages, it's the version referenced by the latest tag. If the latest tag is not set, it's the most recently published package version.

listPackageVersionsResponse_httpStatus :: Lens' ListPackageVersionsResponse Int

The response's http status code.