Network.AWS.CodeArtifact.ListRepositoriesInDomain

Description

Returns a list of RepositorySummary objects. Each RepositorySummary contains information about a repository in the specified domain and that matches the input parameters.

This operation returns paginated results.

Creating a Request

data ListRepositoriesInDomain

See: newListRepositoriesInDomain smart constructor.

Constructors

ListRepositoriesInDomain'

Fields

repositoryPrefix :: Maybe Text

A prefix used to filter returned repositories. Only repositories with names that start with repositoryPrefix are returned.

domainOwner :: Maybe Text

The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

nextToken :: Maybe Text

The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

administratorAccount :: Maybe Text

Filter the list of repositories to only include those that are managed by the AWS account ID.

maxResults :: Maybe Natural

The maximum number of results to return per page.

domain :: Text

The name of the domain that contains the returned list of repositories.

newListRepositoriesInDomain

Arguments

:: Text
$sel:domain:ListRepositoriesInDomain'

-> ListRepositoriesInDomain

Create a value of ListRepositoriesInDomain with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:repositoryPrefix:ListRepositoriesInDomain', listRepositoriesInDomain_repositoryPrefix - A prefix used to filter returned repositories. Only repositories with names that start with repositoryPrefix are returned.

$sel:domainOwner:ListRepositoriesInDomain', listRepositoriesInDomain_domainOwner - The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

$sel:nextToken:ListRepositoriesInDomain', listRepositoriesInDomain_nextToken - The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

$sel:administratorAccount:ListRepositoriesInDomain', listRepositoriesInDomain_administratorAccount - Filter the list of repositories to only include those that are managed by the AWS account ID.

$sel:maxResults:ListRepositoriesInDomain', listRepositoriesInDomain_maxResults - The maximum number of results to return per page.

$sel:domain:ListRepositoriesInDomain', listRepositoriesInDomain_domain - The name of the domain that contains the returned list of repositories.

Request Lenses

listRepositoriesInDomain_repositoryPrefix :: Lens' ListRepositoriesInDomain (Maybe Text)

A prefix used to filter returned repositories. Only repositories with names that start with repositoryPrefix are returned.

listRepositoriesInDomain_domainOwner :: Lens' ListRepositoriesInDomain (Maybe Text)

The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

listRepositoriesInDomain_nextToken :: Lens' ListRepositoriesInDomain (Maybe Text)

The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

listRepositoriesInDomain_administratorAccount :: Lens' ListRepositoriesInDomain (Maybe Text)

Filter the list of repositories to only include those that are managed by the AWS account ID.

listRepositoriesInDomain_maxResults :: Lens' ListRepositoriesInDomain (Maybe Natural)

The maximum number of results to return per page.

listRepositoriesInDomain_domain :: Lens' ListRepositoriesInDomain Text

The name of the domain that contains the returned list of repositories.

Destructuring the Response

data ListRepositoriesInDomainResponse

See: newListRepositoriesInDomainResponse smart constructor.

Constructors

ListRepositoriesInDomainResponse'

Fields

repositories :: Maybe [RepositorySummary]

The returned list of repositories.

nextToken :: Maybe Text

If there are additional results, this is the token for the next set of results.

httpStatus :: Int

The response's http status code.

newListRepositoriesInDomainResponse

Arguments

:: Int
$sel:httpStatus:ListRepositoriesInDomainResponse'

-> ListRepositoriesInDomainResponse

Create a value of ListRepositoriesInDomainResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:repositories:ListRepositoriesInDomainResponse', listRepositoriesInDomainResponse_repositories - The returned list of repositories.

$sel:nextToken:ListRepositoriesInDomain', listRepositoriesInDomainResponse_nextToken - If there are additional results, this is the token for the next set of results.

$sel:httpStatus:ListRepositoriesInDomainResponse', listRepositoriesInDomainResponse_httpStatus - The response's http status code.

Response Lenses

listRepositoriesInDomainResponse_repositories :: Lens' ListRepositoriesInDomainResponse (Maybe [RepositorySummary])

The returned list of repositories.

listRepositoriesInDomainResponse_nextToken :: Lens' ListRepositoriesInDomainResponse (Maybe Text)

If there are additional results, this is the token for the next set of results.

listRepositoriesInDomainResponse_httpStatus :: Lens' ListRepositoriesInDomainResponse Int

The response's http status code.