Network.AWS.CodeArtifact.PutRepositoryPermissionsPolicy

Description

Sets the resource policy on a repository that specifies permissions to access it.

When you call PutRepositoryPermissionsPolicy, the resource policy on the repository is ignored when evaluting permissions. This ensures that the owner of a repository cannot lock themselves out of the repository, which would prevent them from being able to update the resource policy.

Creating a Request

data PutRepositoryPermissionsPolicy

See: newPutRepositoryPermissionsPolicy smart constructor.

Constructors

PutRepositoryPermissionsPolicy'

Fields

policyRevision :: Maybe Text

Sets the revision of the resource policy that specifies permissions to access the repository. This revision is used for optimistic locking, which prevents others from overwriting your changes to the repository's resource policy.

domainOwner :: Maybe Text

The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

domain :: Text

The name of the domain containing the repository to set the resource policy on.

repository :: Text

The name of the repository to set the resource policy on.

policyDocument :: Text

A valid displayable JSON Aspen policy string to be set as the access control resource policy on the provided repository.

newPutRepositoryPermissionsPolicy

Arguments

:: Text
$sel:domain:PutRepositoryPermissionsPolicy'

-> Text
$sel:repository:PutRepositoryPermissionsPolicy'

-> Text
$sel:policyDocument:PutRepositoryPermissionsPolicy'

-> PutRepositoryPermissionsPolicy

Create a value of PutRepositoryPermissionsPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policyRevision:PutRepositoryPermissionsPolicy', putRepositoryPermissionsPolicy_policyRevision - Sets the revision of the resource policy that specifies permissions to access the repository. This revision is used for optimistic locking, which prevents others from overwriting your changes to the repository's resource policy.

$sel:domainOwner:PutRepositoryPermissionsPolicy', putRepositoryPermissionsPolicy_domainOwner - The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

$sel:domain:PutRepositoryPermissionsPolicy', putRepositoryPermissionsPolicy_domain - The name of the domain containing the repository to set the resource policy on.

$sel:repository:PutRepositoryPermissionsPolicy', putRepositoryPermissionsPolicy_repository - The name of the repository to set the resource policy on.

$sel:policyDocument:PutRepositoryPermissionsPolicy', putRepositoryPermissionsPolicy_policyDocument - A valid displayable JSON Aspen policy string to be set as the access control resource policy on the provided repository.

Request Lenses

putRepositoryPermissionsPolicy_policyRevision :: Lens' PutRepositoryPermissionsPolicy (Maybe Text)

Sets the revision of the resource policy that specifies permissions to access the repository. This revision is used for optimistic locking, which prevents others from overwriting your changes to the repository's resource policy.

putRepositoryPermissionsPolicy_domainOwner :: Lens' PutRepositoryPermissionsPolicy (Maybe Text)

The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

putRepositoryPermissionsPolicy_domain :: Lens' PutRepositoryPermissionsPolicy Text

The name of the domain containing the repository to set the resource policy on.

putRepositoryPermissionsPolicy_repository :: Lens' PutRepositoryPermissionsPolicy Text

The name of the repository to set the resource policy on.

putRepositoryPermissionsPolicy_policyDocument :: Lens' PutRepositoryPermissionsPolicy Text

A valid displayable JSON Aspen policy string to be set as the access control resource policy on the provided repository.

Destructuring the Response

data PutRepositoryPermissionsPolicyResponse

See: newPutRepositoryPermissionsPolicyResponse smart constructor.

Constructors

PutRepositoryPermissionsPolicyResponse'

Fields

policy :: Maybe ResourcePolicy

The resource policy that was set after processing the request.

httpStatus :: Int

The response's http status code.

newPutRepositoryPermissionsPolicyResponse

Arguments

:: Int
$sel:httpStatus:PutRepositoryPermissionsPolicyResponse'

-> PutRepositoryPermissionsPolicyResponse

Create a value of PutRepositoryPermissionsPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policy:PutRepositoryPermissionsPolicyResponse', putRepositoryPermissionsPolicyResponse_policy - The resource policy that was set after processing the request.

$sel:httpStatus:PutRepositoryPermissionsPolicyResponse', putRepositoryPermissionsPolicyResponse_httpStatus - The response's http status code.

Response Lenses

putRepositoryPermissionsPolicyResponse_policy :: Lens' PutRepositoryPermissionsPolicyResponse (Maybe ResourcePolicy)

The resource policy that was set after processing the request.

putRepositoryPermissionsPolicyResponse_httpStatus :: Lens' PutRepositoryPermissionsPolicyResponse Int

The response's http status code.