Network.AWS.CodeArtifact.Types

Description

Service Configuration

defaultService :: Service

API version 2018-09-22 of the Amazon CodeArtifact SDK configuration.

Errors

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The operation did not succeed because a parameter in the request was sent with an invalid value.

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

The operation did not succeed because of an unauthorized access attempt.

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

The operation did not succeed because prerequisites are not met.

_ServiceQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The operation did not succeed because it would have exceeded a service limit for your account.

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The operation did not succeed because too many requests are sent to the service.

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

The operation did not succeed because of an error that occurred inside AWS CodeArtifact.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The operation did not succeed because the resource requested is not found in the service.

DomainStatus

newtype DomainStatus

Constructors

DomainStatus'

Fields

fromDomainStatus :: Text

Bundled Patterns

pattern DomainStatus_Active :: DomainStatus

pattern DomainStatus_Deleted :: DomainStatus

ExternalConnectionStatus

newtype ExternalConnectionStatus

Constructors

ExternalConnectionStatus'

Fields

fromExternalConnectionStatus :: Text

Bundled Patterns

pattern ExternalConnectionStatus_Available :: ExternalConnectionStatus

HashAlgorithm

newtype HashAlgorithm

Constructors

HashAlgorithm'

Fields

fromHashAlgorithm :: Text

Bundled Patterns

pattern HashAlgorithm_MD5 :: HashAlgorithm

pattern HashAlgorithm_SHA_1 :: HashAlgorithm

pattern HashAlgorithm_SHA_256 :: HashAlgorithm

pattern HashAlgorithm_SHA_512 :: HashAlgorithm

PackageFormat

newtype PackageFormat

Constructors

PackageFormat'

Fields

fromPackageFormat :: Text

Bundled Patterns

pattern PackageFormat_Maven :: PackageFormat

pattern PackageFormat_Npm :: PackageFormat

pattern PackageFormat_Nuget :: PackageFormat

pattern PackageFormat_Pypi :: PackageFormat

PackageVersionErrorCode

newtype PackageVersionErrorCode

Constructors

PackageVersionErrorCode'

Fields

fromPackageVersionErrorCode :: Text

Bundled Patterns

pattern PackageVersionErrorCode_ALREADY_EXISTS :: PackageVersionErrorCode

pattern PackageVersionErrorCode_MISMATCHED_REVISION :: PackageVersionErrorCode

pattern PackageVersionErrorCode_MISMATCHED_STATUS :: PackageVersionErrorCode

pattern PackageVersionErrorCode_NOT_ALLOWED :: PackageVersionErrorCode

pattern PackageVersionErrorCode_NOT_FOUND :: PackageVersionErrorCode

pattern PackageVersionErrorCode_SKIPPED :: PackageVersionErrorCode

PackageVersionSortType

newtype PackageVersionSortType

Constructors

PackageVersionSortType'

Fields

fromPackageVersionSortType :: Text

Bundled Patterns

pattern PackageVersionSortType_PUBLISHED_TIME :: PackageVersionSortType

PackageVersionStatus

newtype PackageVersionStatus

Constructors

PackageVersionStatus'

Fields

fromPackageVersionStatus :: Text

Bundled Patterns

pattern PackageVersionStatus_Archived :: PackageVersionStatus

pattern PackageVersionStatus_Deleted :: PackageVersionStatus

pattern PackageVersionStatus_Disposed :: PackageVersionStatus

pattern PackageVersionStatus_Published :: PackageVersionStatus

pattern PackageVersionStatus_Unfinished :: PackageVersionStatus

pattern PackageVersionStatus_Unlisted :: PackageVersionStatus

AssetSummary

data AssetSummary

Contains details about a package version asset.

See: newAssetSummary smart constructor.

Constructors

AssetSummary'

Fields

size :: Maybe Integer

The size of the asset.

hashes :: Maybe (HashMap HashAlgorithm Text)

The hashes of the asset.

name :: Text

The name of the asset.

newAssetSummary

Arguments

:: Text
$sel:name:AssetSummary'

-> AssetSummary

Create a value of AssetSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:size:AssetSummary', assetSummary_size - The size of the asset.

$sel:hashes:AssetSummary', assetSummary_hashes - The hashes of the asset.

$sel:name:AssetSummary', assetSummary_name - The name of the asset.

assetSummary_size :: Lens' AssetSummary (Maybe Integer)

The size of the asset.

assetSummary_hashes :: Lens' AssetSummary (Maybe (HashMap HashAlgorithm Text))

The hashes of the asset.

assetSummary_name :: Lens' AssetSummary Text

The name of the asset.

DomainDescription

data DomainDescription

Information about a domain. A domain is a container for repositories. When you create a domain, it is empty until you add one or more repositories.

See: newDomainDescription smart constructor.

Constructors

DomainDescription'

Fields

assetSizeBytes :: Maybe Integer

The total size of all assets in the domain.

status :: Maybe DomainStatus

The current status of a domain. The valid values are

Active

Deleted

arn :: Maybe Text

The Amazon Resource Name (ARN) of the domain.

createdTime :: Maybe POSIX

A timestamp that represents the date and time the domain was created.

s3BucketArn :: Maybe Text

The Amazon Resource Name (ARN) of the Amazon S3 bucket that is used to store package assets in the domain.

owner :: Maybe Text

The AWS account ID that owns the domain.

repositoryCount :: Maybe Int

The number of repositories in the domain.

name :: Maybe Text

The name of the domain.

encryptionKey :: Maybe Text

The ARN of an AWS Key Management Service (AWS KMS) key associated with a domain.

newDomainDescription :: DomainDescription

Create a value of DomainDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:assetSizeBytes:DomainDescription', domainDescription_assetSizeBytes - The total size of all assets in the domain.

$sel:status:DomainDescription', domainDescription_status - The current status of a domain. The valid values are

Active

Deleted

$sel:arn:DomainDescription', domainDescription_arn - The Amazon Resource Name (ARN) of the domain.

$sel:createdTime:DomainDescription', domainDescription_createdTime - A timestamp that represents the date and time the domain was created.

$sel:s3BucketArn:DomainDescription', domainDescription_s3BucketArn - The Amazon Resource Name (ARN) of the Amazon S3 bucket that is used to store package assets in the domain.

$sel:owner:DomainDescription', domainDescription_owner - The AWS account ID that owns the domain.

$sel:repositoryCount:DomainDescription', domainDescription_repositoryCount - The number of repositories in the domain.

$sel:name:DomainDescription', domainDescription_name - The name of the domain.

$sel:encryptionKey:DomainDescription', domainDescription_encryptionKey - The ARN of an AWS Key Management Service (AWS KMS) key associated with a domain.

domainDescription_assetSizeBytes :: Lens' DomainDescription (Maybe Integer)

The total size of all assets in the domain.

domainDescription_status :: Lens' DomainDescription (Maybe DomainStatus)

The current status of a domain. The valid values are

Active

Deleted

domainDescription_arn :: Lens' DomainDescription (Maybe Text)

The Amazon Resource Name (ARN) of the domain.

domainDescription_createdTime :: Lens' DomainDescription (Maybe UTCTime)

A timestamp that represents the date and time the domain was created.

domainDescription_s3BucketArn :: Lens' DomainDescription (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon S3 bucket that is used to store package assets in the domain.

domainDescription_owner :: Lens' DomainDescription (Maybe Text)

The AWS account ID that owns the domain.

domainDescription_repositoryCount :: Lens' DomainDescription (Maybe Int)

The number of repositories in the domain.

domainDescription_name :: Lens' DomainDescription (Maybe Text)

The name of the domain.

domainDescription_encryptionKey :: Lens' DomainDescription (Maybe Text)

The ARN of an AWS Key Management Service (AWS KMS) key associated with a domain.

DomainSummary

data DomainSummary

Information about a domain, including its name, Amazon Resource Name (ARN), and status. The ListDomains operation returns a list of DomainSummary objects.

See: newDomainSummary smart constructor.

Constructors

DomainSummary'

Fields

status :: Maybe DomainStatus

A string that contains the status of the domain. The valid values are:

Active

Deleted

arn :: Maybe Text

The ARN of the domain.

createdTime :: Maybe POSIX

A timestamp that contains the date and time the domain was created.

owner :: Maybe Text

The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

name :: Maybe Text

The name of the domain.

encryptionKey :: Maybe Text

The key used to encrypt the domain.

newDomainSummary :: DomainSummary

Create a value of DomainSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DomainSummary', domainSummary_status - A string that contains the status of the domain. The valid values are:

Active

Deleted

$sel:arn:DomainSummary', domainSummary_arn - The ARN of the domain.

$sel:createdTime:DomainSummary', domainSummary_createdTime - A timestamp that contains the date and time the domain was created.

$sel:owner:DomainSummary', domainSummary_owner - The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

$sel:name:DomainSummary', domainSummary_name - The name of the domain.

$sel:encryptionKey:DomainSummary', domainSummary_encryptionKey - The key used to encrypt the domain.

domainSummary_status :: Lens' DomainSummary (Maybe DomainStatus)

A string that contains the status of the domain. The valid values are:

Active

Deleted

domainSummary_arn :: Lens' DomainSummary (Maybe Text)

The ARN of the domain.

domainSummary_createdTime :: Lens' DomainSummary (Maybe UTCTime)

A timestamp that contains the date and time the domain was created.

domainSummary_owner :: Lens' DomainSummary (Maybe Text)

The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

domainSummary_name :: Lens' DomainSummary (Maybe Text)

The name of the domain.

domainSummary_encryptionKey :: Lens' DomainSummary (Maybe Text)

The key used to encrypt the domain.

LicenseInfo

data LicenseInfo

Details of the license data.

See: newLicenseInfo smart constructor.

Constructors

LicenseInfo'

Fields

url :: Maybe Text

The URL for license data.

name :: Maybe Text

Name of the license.

newLicenseInfo :: LicenseInfo

Create a value of LicenseInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:url:LicenseInfo', licenseInfo_url - The URL for license data.

$sel:name:LicenseInfo', licenseInfo_name - Name of the license.

licenseInfo_url :: Lens' LicenseInfo (Maybe Text)

The URL for license data.

licenseInfo_name :: Lens' LicenseInfo (Maybe Text)

Name of the license.

PackageDependency

data PackageDependency

Details about a package dependency.

See: newPackageDependency smart constructor.

Constructors

PackageDependency'

Fields

dependencyType :: Maybe Text

The type of a package dependency. The possible values depend on the package type. Example types are compile, runtime, and test for Maven packages, and dev, prod, and optional for npm packages.

namespace :: Maybe Text

The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

package :: Maybe Text

The name of the package that this package depends on.

versionRequirement :: Maybe Text

The required version, or version range, of the package that this package depends on. The version format is specific to the package type. For example, the following are possible valid required versions: 1.2.3, ^2.3.4, or 4.x.

newPackageDependency :: PackageDependency

Create a value of PackageDependency with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dependencyType:PackageDependency', packageDependency_dependencyType - The type of a package dependency. The possible values depend on the package type. Example types are compile, runtime, and test for Maven packages, and dev, prod, and optional for npm packages.

$sel:namespace:PackageDependency', packageDependency_namespace - The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

$sel:package:PackageDependency', packageDependency_package - The name of the package that this package depends on.

$sel:versionRequirement:PackageDependency', packageDependency_versionRequirement - The required version, or version range, of the package that this package depends on. The version format is specific to the package type. For example, the following are possible valid required versions: 1.2.3, ^2.3.4, or 4.x.

packageDependency_dependencyType :: Lens' PackageDependency (Maybe Text)

The type of a package dependency. The possible values depend on the package type. Example types are compile, runtime, and test for Maven packages, and dev, prod, and optional for npm packages.

packageDependency_namespace :: Lens' PackageDependency (Maybe Text)

The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

packageDependency_package :: Lens' PackageDependency (Maybe Text)

The name of the package that this package depends on.

packageDependency_versionRequirement :: Lens' PackageDependency (Maybe Text)

The required version, or version range, of the package that this package depends on. The version format is specific to the package type. For example, the following are possible valid required versions: 1.2.3, ^2.3.4, or 4.x.

PackageSummary

data PackageSummary

Details about a package, including its format, namespace, and name. The ListPackages operation returns a list of PackageSummary objects.

See: newPackageSummary smart constructor.

Constructors

PackageSummary'

Fields

format :: Maybe PackageFormat

The format of the package. Valid values are:

npm

pypi

maven

namespace :: Maybe Text

The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

package :: Maybe Text

The name of the package.

newPackageSummary :: PackageSummary

Create a value of PackageSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:format:PackageSummary', packageSummary_format - The format of the package. Valid values are:

npm

pypi

maven

$sel:namespace:PackageSummary', packageSummary_namespace - The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

$sel:package:PackageSummary', packageSummary_package - The name of the package.

packageSummary_format :: Lens' PackageSummary (Maybe PackageFormat)

The format of the package. Valid values are:

npm

pypi

maven

packageSummary_namespace :: Lens' PackageSummary (Maybe Text)

The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

packageSummary_package :: Lens' PackageSummary (Maybe Text)

The name of the package.

PackageVersionDescription

data PackageVersionDescription

Details about a package version.

See: newPackageVersionDescription smart constructor.

Constructors

PackageVersionDescription'

Fields

summary :: Maybe Text

A summary of the package version. The summary is extracted from the package. The information in and detail level of the summary depends on the package version's format.

homePage :: Maybe Text

The homepage associated with the package.

status :: Maybe PackageVersionStatus

A string that contains the status of the package version. It can be one of the following:

Published

Unfinished

Unlisted

Archived

Disposed

sourceCodeRepository :: Maybe Text

The repository for the source code in the package version, or the source code used to build it.

packageName :: Maybe Text

The name of the requested package.

publishedTime :: Maybe POSIX

A timestamp that contains the date and time the package version was published.

format :: Maybe PackageFormat

The format of the package version. The valid package formats are:

npm: A Node Package Manager (npm) package.

pypi: A Python Package Index (PyPI) package.

maven: A Maven package that contains compiled code in a distributable format, such as a JAR file.

namespace :: Maybe Text

The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

version :: Maybe Text

The version of the package.

displayName :: Maybe Text

The name of the package that is displayed. The displayName varies depending on the package version's format. For example, if an npm package is named ui, is in the namespace vue, and has the format npm, then the displayName is @vue/ui.

licenses :: Maybe [LicenseInfo]

Information about licenses associated with the package version.

revision :: Maybe Text

The revision of the package version.

newPackageVersionDescription :: PackageVersionDescription

Create a value of PackageVersionDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:summary:PackageVersionDescription', packageVersionDescription_summary - A summary of the package version. The summary is extracted from the package. The information in and detail level of the summary depends on the package version's format.

$sel:homePage:PackageVersionDescription', packageVersionDescription_homePage - The homepage associated with the package.

$sel:status:PackageVersionDescription', packageVersionDescription_status - A string that contains the status of the package version. It can be one of the following:

Published

Unfinished

Unlisted

Archived

Disposed

$sel:sourceCodeRepository:PackageVersionDescription', packageVersionDescription_sourceCodeRepository - The repository for the source code in the package version, or the source code used to build it.

$sel:packageName:PackageVersionDescription', packageVersionDescription_packageName - The name of the requested package.

$sel:publishedTime:PackageVersionDescription', packageVersionDescription_publishedTime - A timestamp that contains the date and time the package version was published.

$sel:format:PackageVersionDescription', packageVersionDescription_format - The format of the package version. The valid package formats are:

npm: A Node Package Manager (npm) package.

pypi: A Python Package Index (PyPI) package.

maven: A Maven package that contains compiled code in a distributable format, such as a JAR file.

$sel:namespace:PackageVersionDescription', packageVersionDescription_namespace - The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

$sel:version:PackageVersionDescription', packageVersionDescription_version - The version of the package.

$sel:displayName:PackageVersionDescription', packageVersionDescription_displayName - The name of the package that is displayed. The displayName varies depending on the package version's format. For example, if an npm package is named ui, is in the namespace vue, and has the format npm, then the displayName is @vue/ui.

$sel:licenses:PackageVersionDescription', packageVersionDescription_licenses - Information about licenses associated with the package version.

$sel:revision:PackageVersionDescription', packageVersionDescription_revision - The revision of the package version.

packageVersionDescription_summary :: Lens' PackageVersionDescription (Maybe Text)

A summary of the package version. The summary is extracted from the package. The information in and detail level of the summary depends on the package version's format.

packageVersionDescription_homePage :: Lens' PackageVersionDescription (Maybe Text)

The homepage associated with the package.

packageVersionDescription_status :: Lens' PackageVersionDescription (Maybe PackageVersionStatus)

A string that contains the status of the package version. It can be one of the following:

Published

Unfinished

Unlisted

Archived

Disposed

packageVersionDescription_sourceCodeRepository :: Lens' PackageVersionDescription (Maybe Text)

The repository for the source code in the package version, or the source code used to build it.

packageVersionDescription_packageName :: Lens' PackageVersionDescription (Maybe Text)

The name of the requested package.

packageVersionDescription_publishedTime :: Lens' PackageVersionDescription (Maybe UTCTime)

A timestamp that contains the date and time the package version was published.

packageVersionDescription_format :: Lens' PackageVersionDescription (Maybe PackageFormat)

The format of the package version. The valid package formats are:

npm: A Node Package Manager (npm) package.

pypi: A Python Package Index (PyPI) package.

maven: A Maven package that contains compiled code in a distributable format, such as a JAR file.

packageVersionDescription_namespace :: Lens' PackageVersionDescription (Maybe Text)

The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

packageVersionDescription_version :: Lens' PackageVersionDescription (Maybe Text)

The version of the package.

packageVersionDescription_displayName :: Lens' PackageVersionDescription (Maybe Text)

The name of the package that is displayed. The displayName varies depending on the package version's format. For example, if an npm package is named ui, is in the namespace vue, and has the format npm, then the displayName is @vue/ui.

packageVersionDescription_licenses :: Lens' PackageVersionDescription (Maybe [LicenseInfo])

Information about licenses associated with the package version.

packageVersionDescription_revision :: Lens' PackageVersionDescription (Maybe Text)

The revision of the package version.

PackageVersionError

data PackageVersionError

An error associated with package.

See: newPackageVersionError smart constructor.

Constructors

PackageVersionError'

Fields

errorCode :: Maybe PackageVersionErrorCode

The error code associated with the error. Valid error codes are:

ALREADY_EXISTS

MISMATCHED_REVISION

MISMATCHED_STATUS

NOT_ALLOWED

NOT_FOUND

SKIPPED

errorMessage :: Maybe Text

The error message associated with the error.

newPackageVersionError :: PackageVersionError

Create a value of PackageVersionError with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:errorCode:PackageVersionError', packageVersionError_errorCode - The error code associated with the error. Valid error codes are:

ALREADY_EXISTS

MISMATCHED_REVISION

MISMATCHED_STATUS

NOT_ALLOWED

NOT_FOUND

SKIPPED

$sel:errorMessage:PackageVersionError', packageVersionError_errorMessage - The error message associated with the error.

packageVersionError_errorCode :: Lens' PackageVersionError (Maybe PackageVersionErrorCode)

The error code associated with the error. Valid error codes are:

ALREADY_EXISTS

MISMATCHED_REVISION

MISMATCHED_STATUS

NOT_ALLOWED

NOT_FOUND

SKIPPED

packageVersionError_errorMessage :: Lens' PackageVersionError (Maybe Text)

The error message associated with the error.

PackageVersionSummary

data PackageVersionSummary

Details about a package version, including its status, version, and revision. The ListPackageVersions operation returns a list of PackageVersionSummary objects.

See: newPackageVersionSummary smart constructor.

Constructors

PackageVersionSummary'

Fields

revision :: Maybe Text

The revision associated with a package version.

version :: Text

Information about a package version.

status :: PackageVersionStatus

A string that contains the status of the package version. It can be one of the following:

Published

Unfinished

Unlisted

Archived

Disposed

newPackageVersionSummary

Arguments

:: Text
$sel:version:PackageVersionSummary'

-> PackageVersionStatus
$sel:status:PackageVersionSummary'

-> PackageVersionSummary

Create a value of PackageVersionSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:revision:PackageVersionSummary', packageVersionSummary_revision - The revision associated with a package version.

$sel:version:PackageVersionSummary', packageVersionSummary_version - Information about a package version.

$sel:status:PackageVersionSummary', packageVersionSummary_status - A string that contains the status of the package version. It can be one of the following:

Published

Unfinished

Unlisted

Archived

Disposed

packageVersionSummary_revision :: Lens' PackageVersionSummary (Maybe Text)

The revision associated with a package version.

packageVersionSummary_version :: Lens' PackageVersionSummary Text

Information about a package version.

packageVersionSummary_status :: Lens' PackageVersionSummary PackageVersionStatus

A string that contains the status of the package version. It can be one of the following:

Published

Unfinished

Unlisted

Archived

Disposed

RepositoryDescription

data RepositoryDescription

The details of a repository stored in AWS CodeArtifact. A CodeArtifact repository contains a set of package versions, each of which maps to a set of assets. Repositories are polyglot—a single repository can contain packages of any supported type. Each repository exposes endpoints for fetching and publishing packages using tools like the npm CLI, the Maven CLI (mvn), and pip. You can create up to 100 repositories per AWS account.

See: newRepositoryDescription smart constructor.

Constructors

RepositoryDescription'

Fields

upstreams :: Maybe [UpstreamRepositoryInfo]

A list of upstream repositories to associate with the repository. The order of the upstream repositories in the list determines their priority order when AWS CodeArtifact looks for a requested package version. For more information, see Working with upstream repositories.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the repository.

domainOwner :: Maybe Text

The 12-digit account number of the AWS account that owns the domain that contains the repository. It does not include dashes or spaces.

domainName :: Maybe Text

The name of the domain that contains the repository.

name :: Maybe Text

The name of the repository.

administratorAccount :: Maybe Text

The 12-digit account number of the AWS account that manages the repository.

externalConnections :: Maybe [RepositoryExternalConnectionInfo]

An array of external connections associated with the repository.

description :: Maybe Text

A text description of the repository.

newRepositoryDescription :: RepositoryDescription

Create a value of RepositoryDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:upstreams:RepositoryDescription', repositoryDescription_upstreams - A list of upstream repositories to associate with the repository. The order of the upstream repositories in the list determines their priority order when AWS CodeArtifact looks for a requested package version. For more information, see Working with upstream repositories.

$sel:arn:RepositoryDescription', repositoryDescription_arn - The Amazon Resource Name (ARN) of the repository.

$sel:domainOwner:RepositoryDescription', repositoryDescription_domainOwner - The 12-digit account number of the AWS account that owns the domain that contains the repository. It does not include dashes or spaces.

$sel:domainName:RepositoryDescription', repositoryDescription_domainName - The name of the domain that contains the repository.

$sel:name:RepositoryDescription', repositoryDescription_name - The name of the repository.

$sel:administratorAccount:RepositoryDescription', repositoryDescription_administratorAccount - The 12-digit account number of the AWS account that manages the repository.

$sel:externalConnections:RepositoryDescription', repositoryDescription_externalConnections - An array of external connections associated with the repository.

$sel:description:RepositoryDescription', repositoryDescription_description - A text description of the repository.

repositoryDescription_upstreams :: Lens' RepositoryDescription (Maybe [UpstreamRepositoryInfo])

A list of upstream repositories to associate with the repository. The order of the upstream repositories in the list determines their priority order when AWS CodeArtifact looks for a requested package version. For more information, see Working with upstream repositories.

repositoryDescription_arn :: Lens' RepositoryDescription (Maybe Text)

The Amazon Resource Name (ARN) of the repository.

repositoryDescription_domainOwner :: Lens' RepositoryDescription (Maybe Text)

The 12-digit account number of the AWS account that owns the domain that contains the repository. It does not include dashes or spaces.

repositoryDescription_domainName :: Lens' RepositoryDescription (Maybe Text)

The name of the domain that contains the repository.

repositoryDescription_name :: Lens' RepositoryDescription (Maybe Text)

The name of the repository.

repositoryDescription_administratorAccount :: Lens' RepositoryDescription (Maybe Text)

The 12-digit account number of the AWS account that manages the repository.

repositoryDescription_externalConnections :: Lens' RepositoryDescription (Maybe [RepositoryExternalConnectionInfo])

An array of external connections associated with the repository.

repositoryDescription_description :: Lens' RepositoryDescription (Maybe Text)

A text description of the repository.

RepositoryExternalConnectionInfo

data RepositoryExternalConnectionInfo

Contains information about the external connection of a repository.

See: newRepositoryExternalConnectionInfo smart constructor.

Constructors

RepositoryExternalConnectionInfo'

Fields

externalConnectionName :: Maybe Text

The name of the external connection associated with a repository.

status :: Maybe ExternalConnectionStatus

The status of the external connection of a repository. There is one valid value, Available.

packageFormat :: Maybe PackageFormat

The package format associated with a repository's external connection. The valid package formats are:

npm: A Node Package Manager (npm) package.

pypi: A Python Package Index (PyPI) package.

maven: A Maven package that contains compiled code in a distributable format, such as a JAR file.

newRepositoryExternalConnectionInfo :: RepositoryExternalConnectionInfo

Create a value of RepositoryExternalConnectionInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:externalConnectionName:RepositoryExternalConnectionInfo', repositoryExternalConnectionInfo_externalConnectionName - The name of the external connection associated with a repository.

$sel:status:RepositoryExternalConnectionInfo', repositoryExternalConnectionInfo_status - The status of the external connection of a repository. There is one valid value, Available.

$sel:packageFormat:RepositoryExternalConnectionInfo', repositoryExternalConnectionInfo_packageFormat - The package format associated with a repository's external connection. The valid package formats are:

npm: A Node Package Manager (npm) package.

pypi: A Python Package Index (PyPI) package.

maven: A Maven package that contains compiled code in a distributable format, such as a JAR file.

repositoryExternalConnectionInfo_externalConnectionName :: Lens' RepositoryExternalConnectionInfo (Maybe Text)

The name of the external connection associated with a repository.

repositoryExternalConnectionInfo_status :: Lens' RepositoryExternalConnectionInfo (Maybe ExternalConnectionStatus)

The status of the external connection of a repository. There is one valid value, Available.

repositoryExternalConnectionInfo_packageFormat :: Lens' RepositoryExternalConnectionInfo (Maybe PackageFormat)

The package format associated with a repository's external connection. The valid package formats are:

npm: A Node Package Manager (npm) package.

pypi: A Python Package Index (PyPI) package.

maven: A Maven package that contains compiled code in a distributable format, such as a JAR file.

RepositorySummary

data RepositorySummary

Details about a repository, including its Amazon Resource Name (ARN), description, and domain information. The ListRepositories operation returns a list of RepositorySummary objects.

See: newRepositorySummary smart constructor.

Constructors

RepositorySummary'

Fields

arn :: Maybe Text

The ARN of the repository.

domainOwner :: Maybe Text

The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

domainName :: Maybe Text

The name of the domain that contains the repository.

name :: Maybe Text

The name of the repository.

administratorAccount :: Maybe Text

The AWS account ID that manages the repository.

description :: Maybe Text

The description of the repository.

newRepositorySummary :: RepositorySummary

Create a value of RepositorySummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:RepositorySummary', repositorySummary_arn - The ARN of the repository.

$sel:domainOwner:RepositorySummary', repositorySummary_domainOwner - The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

$sel:domainName:RepositorySummary', repositorySummary_domainName - The name of the domain that contains the repository.

$sel:name:RepositorySummary', repositorySummary_name - The name of the repository.

$sel:administratorAccount:RepositorySummary', repositorySummary_administratorAccount - The AWS account ID that manages the repository.

$sel:description:RepositorySummary', repositorySummary_description - The description of the repository.

repositorySummary_arn :: Lens' RepositorySummary (Maybe Text)

The ARN of the repository.

repositorySummary_domainOwner :: Lens' RepositorySummary (Maybe Text)

The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

repositorySummary_domainName :: Lens' RepositorySummary (Maybe Text)

The name of the domain that contains the repository.

repositorySummary_name :: Lens' RepositorySummary (Maybe Text)

The name of the repository.

repositorySummary_administratorAccount :: Lens' RepositorySummary (Maybe Text)

The AWS account ID that manages the repository.

repositorySummary_description :: Lens' RepositorySummary (Maybe Text)

The description of the repository.

ResourcePolicy

data ResourcePolicy

An AWS CodeArtifact resource policy that contains a resource ARN, document details, and a revision.

See: newResourcePolicy smart constructor.

Constructors

ResourcePolicy'

Fields

resourceArn :: Maybe Text

The ARN of the resource associated with the resource policy

document :: Maybe Text

The resource policy formatted in JSON.

revision :: Maybe Text

The current revision of the resource policy.

newResourcePolicy :: ResourcePolicy

Create a value of ResourcePolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:ResourcePolicy', resourcePolicy_resourceArn - The ARN of the resource associated with the resource policy

$sel:document:ResourcePolicy', resourcePolicy_document - The resource policy formatted in JSON.

$sel:revision:ResourcePolicy', resourcePolicy_revision - The current revision of the resource policy.

resourcePolicy_resourceArn :: Lens' ResourcePolicy (Maybe Text)

The ARN of the resource associated with the resource policy

resourcePolicy_document :: Lens' ResourcePolicy (Maybe Text)

The resource policy formatted in JSON.

resourcePolicy_revision :: Lens' ResourcePolicy (Maybe Text)

The current revision of the resource policy.

SuccessfulPackageVersionInfo

data SuccessfulPackageVersionInfo

Contains the revision and status of a package version.

See: newSuccessfulPackageVersionInfo smart constructor.

Constructors

SuccessfulPackageVersionInfo'

Fields

status :: Maybe PackageVersionStatus

The status of a package version. Valid statuses are:

Published

Unfinished

Unlisted

Archived

Disposed

revision :: Maybe Text

The revision of a package version.

newSuccessfulPackageVersionInfo :: SuccessfulPackageVersionInfo

Create a value of SuccessfulPackageVersionInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:SuccessfulPackageVersionInfo', successfulPackageVersionInfo_status - The status of a package version. Valid statuses are:

Published

Unfinished

Unlisted

Archived

Disposed

$sel:revision:SuccessfulPackageVersionInfo', successfulPackageVersionInfo_revision - The revision of a package version.

successfulPackageVersionInfo_status :: Lens' SuccessfulPackageVersionInfo (Maybe PackageVersionStatus)

The status of a package version. Valid statuses are:

Published

Unfinished

Unlisted

Archived

Disposed

successfulPackageVersionInfo_revision :: Lens' SuccessfulPackageVersionInfo (Maybe Text)

The revision of a package version.

Tag

data Tag

A tag is a key-value pair that can be used to manage, search for, or filter resources in AWS CodeArtifact.

See: newTag smart constructor.

Constructors

Tag'

Fields

key :: Text

The tag key.

value :: Text

The tag value.

newTag

Arguments

:: Text
$sel:key:Tag'

-> Text
$sel:value:Tag'

-> Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:Tag', tag_key - The tag key.

$sel:value:Tag', tag_value - The tag value.

tag_key :: Lens' Tag Text

The tag key.

tag_value :: Lens' Tag Text

The tag value.

UpstreamRepository

data UpstreamRepository

Information about an upstream repository. A list of UpstreamRepository objects is an input parameter to CreateRepository and UpdateRepository.

See: newUpstreamRepository smart constructor.

Constructors

UpstreamRepository'

Fields

repositoryName :: Text

The name of an upstream repository.

newUpstreamRepository

Arguments

:: Text
$sel:repositoryName:UpstreamRepository'

-> UpstreamRepository

Create a value of UpstreamRepository with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:repositoryName:UpstreamRepository', upstreamRepository_repositoryName - The name of an upstream repository.

upstreamRepository_repositoryName :: Lens' UpstreamRepository Text

The name of an upstream repository.

UpstreamRepositoryInfo

data UpstreamRepositoryInfo

Information about an upstream repository.

See: newUpstreamRepositoryInfo smart constructor.

Constructors

UpstreamRepositoryInfo'

Fields

repositoryName :: Maybe Text

The name of an upstream repository.

newUpstreamRepositoryInfo :: UpstreamRepositoryInfo

Create a value of UpstreamRepositoryInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:repositoryName:UpstreamRepositoryInfo', upstreamRepositoryInfo_repositoryName - The name of an upstream repository.

upstreamRepositoryInfo_repositoryName :: Lens' UpstreamRepositoryInfo (Maybe Text)

The name of an upstream repository.