Network.AWS.CodeArtifact.Types.DomainDescription

Description

Documentation

data DomainDescription

Information about a domain. A domain is a container for repositories. When you create a domain, it is empty until you add one or more repositories.

See: newDomainDescription smart constructor.

Constructors

DomainDescription'

Fields

assetSizeBytes :: Maybe Integer

The total size of all assets in the domain.

status :: Maybe DomainStatus

The current status of a domain. The valid values are

Active

Deleted

arn :: Maybe Text

The Amazon Resource Name (ARN) of the domain.

createdTime :: Maybe POSIX

A timestamp that represents the date and time the domain was created.

s3BucketArn :: Maybe Text

The Amazon Resource Name (ARN) of the Amazon S3 bucket that is used to store package assets in the domain.

owner :: Maybe Text

The AWS account ID that owns the domain.

repositoryCount :: Maybe Int

The number of repositories in the domain.

name :: Maybe Text

The name of the domain.

encryptionKey :: Maybe Text

The ARN of an AWS Key Management Service (AWS KMS) key associated with a domain.

newDomainDescription :: DomainDescription

Create a value of DomainDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:assetSizeBytes:DomainDescription', domainDescription_assetSizeBytes - The total size of all assets in the domain.

$sel:status:DomainDescription', domainDescription_status - The current status of a domain. The valid values are

Active

Deleted

$sel:arn:DomainDescription', domainDescription_arn - The Amazon Resource Name (ARN) of the domain.

$sel:createdTime:DomainDescription', domainDescription_createdTime - A timestamp that represents the date and time the domain was created.

$sel:s3BucketArn:DomainDescription', domainDescription_s3BucketArn - The Amazon Resource Name (ARN) of the Amazon S3 bucket that is used to store package assets in the domain.

$sel:owner:DomainDescription', domainDescription_owner - The AWS account ID that owns the domain.

$sel:repositoryCount:DomainDescription', domainDescription_repositoryCount - The number of repositories in the domain.

$sel:name:DomainDescription', domainDescription_name - The name of the domain.

$sel:encryptionKey:DomainDescription', domainDescription_encryptionKey - The ARN of an AWS Key Management Service (AWS KMS) key associated with a domain.

domainDescription_assetSizeBytes :: Lens' DomainDescription (Maybe Integer)

The total size of all assets in the domain.

domainDescription_status :: Lens' DomainDescription (Maybe DomainStatus)

The current status of a domain. The valid values are

Active

Deleted

domainDescription_arn :: Lens' DomainDescription (Maybe Text)

The Amazon Resource Name (ARN) of the domain.

domainDescription_createdTime :: Lens' DomainDescription (Maybe UTCTime)

A timestamp that represents the date and time the domain was created.

domainDescription_s3BucketArn :: Lens' DomainDescription (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon S3 bucket that is used to store package assets in the domain.

domainDescription_owner :: Lens' DomainDescription (Maybe Text)

The AWS account ID that owns the domain.

domainDescription_repositoryCount :: Lens' DomainDescription (Maybe Int)

The number of repositories in the domain.

domainDescription_name :: Lens' DomainDescription (Maybe Text)

The name of the domain.

domainDescription_encryptionKey :: Lens' DomainDescription (Maybe Text)

The ARN of an AWS Key Management Service (AWS KMS) key associated with a domain.