Network.AWS.CodeArtifact.Types.RepositoryDescription

Description

Documentation

data RepositoryDescription

The details of a repository stored in AWS CodeArtifact. A CodeArtifact repository contains a set of package versions, each of which maps to a set of assets. Repositories are polyglot—a single repository can contain packages of any supported type. Each repository exposes endpoints for fetching and publishing packages using tools like the npm CLI, the Maven CLI (mvn), and pip. You can create up to 100 repositories per AWS account.

See: newRepositoryDescription smart constructor.

Constructors

RepositoryDescription'

Fields

upstreams :: Maybe [UpstreamRepositoryInfo]

A list of upstream repositories to associate with the repository. The order of the upstream repositories in the list determines their priority order when AWS CodeArtifact looks for a requested package version. For more information, see Working with upstream repositories.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the repository.

domainOwner :: Maybe Text

The 12-digit account number of the AWS account that owns the domain that contains the repository. It does not include dashes or spaces.

domainName :: Maybe Text

The name of the domain that contains the repository.

name :: Maybe Text

The name of the repository.

administratorAccount :: Maybe Text

The 12-digit account number of the AWS account that manages the repository.

externalConnections :: Maybe [RepositoryExternalConnectionInfo]

An array of external connections associated with the repository.

description :: Maybe Text

A text description of the repository.

newRepositoryDescription :: RepositoryDescription

Create a value of RepositoryDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:upstreams:RepositoryDescription', repositoryDescription_upstreams - A list of upstream repositories to associate with the repository. The order of the upstream repositories in the list determines their priority order when AWS CodeArtifact looks for a requested package version. For more information, see Working with upstream repositories.

$sel:arn:RepositoryDescription', repositoryDescription_arn - The Amazon Resource Name (ARN) of the repository.

$sel:domainOwner:RepositoryDescription', repositoryDescription_domainOwner - The 12-digit account number of the AWS account that owns the domain that contains the repository. It does not include dashes or spaces.

$sel:domainName:RepositoryDescription', repositoryDescription_domainName - The name of the domain that contains the repository.

$sel:name:RepositoryDescription', repositoryDescription_name - The name of the repository.

$sel:administratorAccount:RepositoryDescription', repositoryDescription_administratorAccount - The 12-digit account number of the AWS account that manages the repository.

$sel:externalConnections:RepositoryDescription', repositoryDescription_externalConnections - An array of external connections associated with the repository.

$sel:description:RepositoryDescription', repositoryDescription_description - A text description of the repository.

repositoryDescription_upstreams :: Lens' RepositoryDescription (Maybe [UpstreamRepositoryInfo])

A list of upstream repositories to associate with the repository. The order of the upstream repositories in the list determines their priority order when AWS CodeArtifact looks for a requested package version. For more information, see Working with upstream repositories.

repositoryDescription_arn :: Lens' RepositoryDescription (Maybe Text)

The Amazon Resource Name (ARN) of the repository.

repositoryDescription_domainOwner :: Lens' RepositoryDescription (Maybe Text)

The 12-digit account number of the AWS account that owns the domain that contains the repository. It does not include dashes or spaces.

repositoryDescription_domainName :: Lens' RepositoryDescription (Maybe Text)

The name of the domain that contains the repository.

repositoryDescription_name :: Lens' RepositoryDescription (Maybe Text)

The name of the repository.

repositoryDescription_administratorAccount :: Lens' RepositoryDescription (Maybe Text)

The 12-digit account number of the AWS account that manages the repository.

repositoryDescription_externalConnections :: Lens' RepositoryDescription (Maybe [RepositoryExternalConnectionInfo])

An array of external connections associated with the repository.

repositoryDescription_description :: Lens' RepositoryDescription (Maybe Text)

A text description of the repository.