Network.AWS.CodeArtifact.Types.RepositorySummary

Description

Documentation

data RepositorySummary

Details about a repository, including its Amazon Resource Name (ARN), description, and domain information. The ListRepositories operation returns a list of RepositorySummary objects.

See: newRepositorySummary smart constructor.

Constructors

RepositorySummary'

Fields

arn :: Maybe Text

The ARN of the repository.

domainOwner :: Maybe Text

The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

domainName :: Maybe Text

The name of the domain that contains the repository.

name :: Maybe Text

The name of the repository.

administratorAccount :: Maybe Text

The AWS account ID that manages the repository.

description :: Maybe Text

The description of the repository.

newRepositorySummary :: RepositorySummary

Create a value of RepositorySummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:RepositorySummary', repositorySummary_arn - The ARN of the repository.

$sel:domainOwner:RepositorySummary', repositorySummary_domainOwner - The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

$sel:domainName:RepositorySummary', repositorySummary_domainName - The name of the domain that contains the repository.

$sel:name:RepositorySummary', repositorySummary_name - The name of the repository.

$sel:administratorAccount:RepositorySummary', repositorySummary_administratorAccount - The AWS account ID that manages the repository.

$sel:description:RepositorySummary', repositorySummary_description - The description of the repository.

repositorySummary_arn :: Lens' RepositorySummary (Maybe Text)

The ARN of the repository.

repositorySummary_domainOwner :: Lens' RepositorySummary (Maybe Text)

The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

repositorySummary_domainName :: Lens' RepositorySummary (Maybe Text)

The name of the domain that contains the repository.

repositorySummary_name :: Lens' RepositorySummary (Maybe Text)

The name of the repository.

repositorySummary_administratorAccount :: Lens' RepositorySummary (Maybe Text)

The AWS account ID that manages the repository.

repositorySummary_description :: Lens' RepositorySummary (Maybe Text)

The description of the repository.