Network.AWS.CodeArtifact.UpdatePackageVersionsStatus

Description

Updates the status of one or more versions of a package.

Creating a Request

data UpdatePackageVersionsStatus

See: newUpdatePackageVersionsStatus smart constructor.

Constructors

UpdatePackageVersionsStatus'

Fields

expectedStatus :: Maybe PackageVersionStatus

The package version’s expected status before it is updated. If expectedStatus is provided, the package version's status is updated only if its status at the time UpdatePackageVersionsStatus is called matches expectedStatus.

versionRevisions :: Maybe (HashMap Text Text)

A map of package versions and package version revisions. The map key is the package version (for example, 3.5.2), and the map value is the package version revision.

namespace :: Maybe Text

The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

domainOwner :: Maybe Text

The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

domain :: Text

The name of the domain that contains the repository that contains the package versions with a status to be updated.

repository :: Text

The repository that contains the package versions with the status you want to update.

format :: PackageFormat

A format that specifies the type of the package with the statuses to update. The valid values are:

npm

pypi

maven

package :: Text

The name of the package with the version statuses to update.

versions :: [Text]

An array of strings that specify the versions of the package with the statuses to update.

targetStatus :: PackageVersionStatus

The status you want to change the package version status to.

newUpdatePackageVersionsStatus

Arguments

:: Text
$sel:domain:UpdatePackageVersionsStatus'

-> Text
$sel:repository:UpdatePackageVersionsStatus'

-> PackageFormat
$sel:format:UpdatePackageVersionsStatus'

-> Text
$sel:package:UpdatePackageVersionsStatus'

-> PackageVersionStatus
$sel:targetStatus:UpdatePackageVersionsStatus'

-> UpdatePackageVersionsStatus

Create a value of UpdatePackageVersionsStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expectedStatus:UpdatePackageVersionsStatus', updatePackageVersionsStatus_expectedStatus - The package version’s expected status before it is updated. If expectedStatus is provided, the package version's status is updated only if its status at the time UpdatePackageVersionsStatus is called matches expectedStatus.

$sel:versionRevisions:UpdatePackageVersionsStatus', updatePackageVersionsStatus_versionRevisions - A map of package versions and package version revisions. The map key is the package version (for example, 3.5.2), and the map value is the package version revision.

$sel:namespace:UpdatePackageVersionsStatus', updatePackageVersionsStatus_namespace - The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

$sel:domainOwner:UpdatePackageVersionsStatus', updatePackageVersionsStatus_domainOwner - The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

$sel:domain:UpdatePackageVersionsStatus', updatePackageVersionsStatus_domain - The name of the domain that contains the repository that contains the package versions with a status to be updated.

$sel:repository:UpdatePackageVersionsStatus', updatePackageVersionsStatus_repository - The repository that contains the package versions with the status you want to update.

$sel:format:UpdatePackageVersionsStatus', updatePackageVersionsStatus_format - A format that specifies the type of the package with the statuses to update. The valid values are:

npm

pypi

maven

$sel:package:UpdatePackageVersionsStatus', updatePackageVersionsStatus_package - The name of the package with the version statuses to update.

$sel:versions:UpdatePackageVersionsStatus', updatePackageVersionsStatus_versions - An array of strings that specify the versions of the package with the statuses to update.

$sel:targetStatus:UpdatePackageVersionsStatus', updatePackageVersionsStatus_targetStatus - The status you want to change the package version status to.

Request Lenses

updatePackageVersionsStatus_expectedStatus :: Lens' UpdatePackageVersionsStatus (Maybe PackageVersionStatus)

The package version’s expected status before it is updated. If expectedStatus is provided, the package version's status is updated only if its status at the time UpdatePackageVersionsStatus is called matches expectedStatus.

updatePackageVersionsStatus_versionRevisions :: Lens' UpdatePackageVersionsStatus (Maybe (HashMap Text Text))

A map of package versions and package version revisions. The map key is the package version (for example, 3.5.2), and the map value is the package version revision.

updatePackageVersionsStatus_namespace :: Lens' UpdatePackageVersionsStatus (Maybe Text)

The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

updatePackageVersionsStatus_domainOwner :: Lens' UpdatePackageVersionsStatus (Maybe Text)

The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

updatePackageVersionsStatus_domain :: Lens' UpdatePackageVersionsStatus Text

The name of the domain that contains the repository that contains the package versions with a status to be updated.

updatePackageVersionsStatus_repository :: Lens' UpdatePackageVersionsStatus Text

The repository that contains the package versions with the status you want to update.

updatePackageVersionsStatus_format :: Lens' UpdatePackageVersionsStatus PackageFormat

A format that specifies the type of the package with the statuses to update. The valid values are:

npm

pypi

maven

updatePackageVersionsStatus_package :: Lens' UpdatePackageVersionsStatus Text

The name of the package with the version statuses to update.

updatePackageVersionsStatus_versions :: Lens' UpdatePackageVersionsStatus [Text]

An array of strings that specify the versions of the package with the statuses to update.

updatePackageVersionsStatus_targetStatus :: Lens' UpdatePackageVersionsStatus PackageVersionStatus

The status you want to change the package version status to.

Destructuring the Response

data UpdatePackageVersionsStatusResponse

See: newUpdatePackageVersionsStatusResponse smart constructor.

Constructors

UpdatePackageVersionsStatusResponse'

Fields

failedVersions :: Maybe (HashMap Text PackageVersionError)

A list of SuccessfulPackageVersionInfo objects, one for each package version with a status that successfully updated.

successfulVersions :: Maybe (HashMap Text SuccessfulPackageVersionInfo)

A list of PackageVersionError objects, one for each package version with a status that failed to update.

httpStatus :: Int

The response's http status code.

newUpdatePackageVersionsStatusResponse

Arguments

:: Int
$sel:httpStatus:UpdatePackageVersionsStatusResponse'

-> UpdatePackageVersionsStatusResponse

Create a value of UpdatePackageVersionsStatusResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failedVersions:UpdatePackageVersionsStatusResponse', updatePackageVersionsStatusResponse_failedVersions - A list of SuccessfulPackageVersionInfo objects, one for each package version with a status that successfully updated.

$sel:successfulVersions:UpdatePackageVersionsStatusResponse', updatePackageVersionsStatusResponse_successfulVersions - A list of PackageVersionError objects, one for each package version with a status that failed to update.

$sel:httpStatus:UpdatePackageVersionsStatusResponse', updatePackageVersionsStatusResponse_httpStatus - The response's http status code.

Response Lenses

updatePackageVersionsStatusResponse_failedVersions :: Lens' UpdatePackageVersionsStatusResponse (Maybe (HashMap Text PackageVersionError))

A list of SuccessfulPackageVersionInfo objects, one for each package version with a status that successfully updated.

updatePackageVersionsStatusResponse_successfulVersions :: Lens' UpdatePackageVersionsStatusResponse (Maybe (HashMap Text SuccessfulPackageVersionInfo))

A list of PackageVersionError objects, one for each package version with a status that failed to update.

updatePackageVersionsStatusResponse_httpStatus :: Lens' UpdatePackageVersionsStatusResponse Int

The response's http status code.