Network.AWS.CodeBuild.CreateProject

Description

Creates a build project.

Creating a Request

data CreateProject

See: newCreateProject smart constructor.

Constructors

CreateProject'

Fields

vpcConfig :: Maybe VpcConfig

VpcConfig enables CodeBuild to access resources in an Amazon VPC.

sourceVersion :: Maybe Text

A version of the build input to be built for this project. If not specified, the latest version is used. If specified, it must be one of:

For CodeCommit: the commit ID, branch, or Git tag to use.

For GitHub: the commit ID, pull request ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a pull request ID is specified, it must use the format pr/pull-request-ID (for example pr/25). If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

For Bitbucket: the commit ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

For Amazon S3: the version ID of the object that represents the build input ZIP file to use.

If sourceVersion is specified at the build level, then that version takes precedence over this sourceVersion (at the project level).

For more information, see Source Version Sample with CodeBuild in the CodeBuild User Guide.

secondaryArtifacts :: Maybe [ProjectArtifacts]

An array of ProjectArtifacts objects.

cache :: Maybe ProjectCache

Stores recently used information so that it can be quickly accessed at a later time.

secondarySourceVersions :: Maybe [ProjectSourceVersion]

An array of ProjectSourceVersion objects. If secondarySourceVersions is specified at the build level, then they take precedence over these secondarySourceVersions (at the project level).

badgeEnabled :: Maybe Bool

Set this to true to generate a publicly accessible URL for your project's build badge.

concurrentBuildLimit :: Maybe Int

The maximum number of concurrent builds that are allowed for this project.

New builds are only started if the current number of builds is less than or equal to this limit. If the current build count meets this limit, new builds are throttled and are not run.

encryptionKey :: Maybe Text

The Key Management Service customer master key (CMK) to be used for encrypting the build output artifacts.

You can use a cross-account KMS key to encrypt the build output artifacts if your service role has permission to that key.

You can specify either the Amazon Resource Name (ARN) of the CMK or, if available, the CMK's alias (using the format alias/<alias-name>).

logsConfig :: Maybe LogsConfig

Information about logs for the build project. These can be logs in CloudWatch Logs, logs uploaded to a specified S3 bucket, or both.

queuedTimeoutInMinutes :: Maybe Natural

The number of minutes a build is allowed to be queued before it times out.

secondarySources :: Maybe [ProjectSource]

An array of ProjectSource objects.

timeoutInMinutes :: Maybe Natural

How long, in minutes, from 5 to 480 (8 hours), for CodeBuild to wait before it times out any build that has not been marked as completed. The default is 60 minutes.

tags :: Maybe [Tag]

A list of tag key and value pairs associated with this build project.

These tags are available for use by Amazon Web Services services that support CodeBuild build project tags.

description :: Maybe Text

A description that makes the build project easy to identify.

buildBatchConfig :: Maybe ProjectBuildBatchConfig

A ProjectBuildBatchConfig object that defines the batch build options for the project.

fileSystemLocations :: Maybe [ProjectFileSystemLocation]

An array of ProjectFileSystemLocation objects for a CodeBuild build project. A ProjectFileSystemLocation object specifies the identifier, location, mountOptions, mountPoint, and type of a file system created using Amazon Elastic File System.

name :: Text

The name of the build project.

source :: ProjectSource

Information about the build input source code for the build project.

artifacts :: ProjectArtifacts

Information about the build output artifacts for the build project.

environment :: ProjectEnvironment

Information about the build environment for the build project.

serviceRole :: Text

The ARN of the IAM role that enables CodeBuild to interact with dependent Amazon Web Services services on behalf of the Amazon Web Services account.

newCreateProject

Arguments

:: Text
$sel:name:CreateProject'

-> ProjectSource
$sel:source:CreateProject'

-> ProjectArtifacts
$sel:artifacts:CreateProject'

-> ProjectEnvironment
$sel:environment:CreateProject'

-> Text
$sel:serviceRole:CreateProject'

-> CreateProject

Create a value of CreateProject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpcConfig:CreateProject', createProject_vpcConfig - VpcConfig enables CodeBuild to access resources in an Amazon VPC.

$sel:sourceVersion:CreateProject', createProject_sourceVersion - A version of the build input to be built for this project. If not specified, the latest version is used. If specified, it must be one of:

For CodeCommit: the commit ID, branch, or Git tag to use.

For GitHub: the commit ID, pull request ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a pull request ID is specified, it must use the format pr/pull-request-ID (for example pr/25). If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

For Bitbucket: the commit ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

For Amazon S3: the version ID of the object that represents the build input ZIP file to use.

If sourceVersion is specified at the build level, then that version takes precedence over this sourceVersion (at the project level).

For more information, see Source Version Sample with CodeBuild in the CodeBuild User Guide.

$sel:secondaryArtifacts:CreateProject', createProject_secondaryArtifacts - An array of ProjectArtifacts objects.

$sel:cache:CreateProject', createProject_cache - Stores recently used information so that it can be quickly accessed at a later time.

$sel:secondarySourceVersions:CreateProject', createProject_secondarySourceVersions - An array of ProjectSourceVersion objects. If secondarySourceVersions is specified at the build level, then they take precedence over these secondarySourceVersions (at the project level).

$sel:badgeEnabled:CreateProject', createProject_badgeEnabled - Set this to true to generate a publicly accessible URL for your project's build badge.

$sel:concurrentBuildLimit:CreateProject', createProject_concurrentBuildLimit - The maximum number of concurrent builds that are allowed for this project.

New builds are only started if the current number of builds is less than or equal to this limit. If the current build count meets this limit, new builds are throttled and are not run.

$sel:encryptionKey:CreateProject', createProject_encryptionKey - The Key Management Service customer master key (CMK) to be used for encrypting the build output artifacts.

You can use a cross-account KMS key to encrypt the build output artifacts if your service role has permission to that key.

You can specify either the Amazon Resource Name (ARN) of the CMK or, if available, the CMK's alias (using the format alias/<alias-name>).

$sel:logsConfig:CreateProject', createProject_logsConfig - Information about logs for the build project. These can be logs in CloudWatch Logs, logs uploaded to a specified S3 bucket, or both.

$sel:queuedTimeoutInMinutes:CreateProject', createProject_queuedTimeoutInMinutes - The number of minutes a build is allowed to be queued before it times out.

$sel:secondarySources:CreateProject', createProject_secondarySources - An array of ProjectSource objects.

$sel:timeoutInMinutes:CreateProject', createProject_timeoutInMinutes - How long, in minutes, from 5 to 480 (8 hours), for CodeBuild to wait before it times out any build that has not been marked as completed. The default is 60 minutes.

$sel:tags:CreateProject', createProject_tags - A list of tag key and value pairs associated with this build project.

These tags are available for use by Amazon Web Services services that support CodeBuild build project tags.

$sel:description:CreateProject', createProject_description - A description that makes the build project easy to identify.

$sel:buildBatchConfig:CreateProject', createProject_buildBatchConfig - A ProjectBuildBatchConfig object that defines the batch build options for the project.

$sel:fileSystemLocations:CreateProject', createProject_fileSystemLocations - An array of ProjectFileSystemLocation objects for a CodeBuild build project. A ProjectFileSystemLocation object specifies the identifier, location, mountOptions, mountPoint, and type of a file system created using Amazon Elastic File System.

$sel:name:CreateProject', createProject_name - The name of the build project.

$sel:source:CreateProject', createProject_source - Information about the build input source code for the build project.

$sel:artifacts:CreateProject', createProject_artifacts - Information about the build output artifacts for the build project.

$sel:environment:CreateProject', createProject_environment - Information about the build environment for the build project.

$sel:serviceRole:CreateProject', createProject_serviceRole - The ARN of the IAM role that enables CodeBuild to interact with dependent Amazon Web Services services on behalf of the Amazon Web Services account.

Request Lenses

createProject_vpcConfig :: Lens' CreateProject (Maybe VpcConfig)

VpcConfig enables CodeBuild to access resources in an Amazon VPC.

createProject_sourceVersion :: Lens' CreateProject (Maybe Text)

A version of the build input to be built for this project. If not specified, the latest version is used. If specified, it must be one of:

For CodeCommit: the commit ID, branch, or Git tag to use.

For GitHub: the commit ID, pull request ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a pull request ID is specified, it must use the format pr/pull-request-ID (for example pr/25). If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

For Bitbucket: the commit ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

For Amazon S3: the version ID of the object that represents the build input ZIP file to use.

If sourceVersion is specified at the build level, then that version takes precedence over this sourceVersion (at the project level).

For more information, see Source Version Sample with CodeBuild in the CodeBuild User Guide.

createProject_secondaryArtifacts :: Lens' CreateProject (Maybe [ProjectArtifacts])

An array of ProjectArtifacts objects.

createProject_cache :: Lens' CreateProject (Maybe ProjectCache)

Stores recently used information so that it can be quickly accessed at a later time.

createProject_secondarySourceVersions :: Lens' CreateProject (Maybe [ProjectSourceVersion])

An array of ProjectSourceVersion objects. If secondarySourceVersions is specified at the build level, then they take precedence over these secondarySourceVersions (at the project level).

createProject_badgeEnabled :: Lens' CreateProject (Maybe Bool)

Set this to true to generate a publicly accessible URL for your project's build badge.

createProject_concurrentBuildLimit :: Lens' CreateProject (Maybe Int)

The maximum number of concurrent builds that are allowed for this project.

New builds are only started if the current number of builds is less than or equal to this limit. If the current build count meets this limit, new builds are throttled and are not run.

createProject_encryptionKey :: Lens' CreateProject (Maybe Text)

The Key Management Service customer master key (CMK) to be used for encrypting the build output artifacts.

You can use a cross-account KMS key to encrypt the build output artifacts if your service role has permission to that key.

You can specify either the Amazon Resource Name (ARN) of the CMK or, if available, the CMK's alias (using the format alias/<alias-name>).

createProject_logsConfig :: Lens' CreateProject (Maybe LogsConfig)

Information about logs for the build project. These can be logs in CloudWatch Logs, logs uploaded to a specified S3 bucket, or both.

createProject_queuedTimeoutInMinutes :: Lens' CreateProject (Maybe Natural)

The number of minutes a build is allowed to be queued before it times out.

createProject_secondarySources :: Lens' CreateProject (Maybe [ProjectSource])

An array of ProjectSource objects.

createProject_timeoutInMinutes :: Lens' CreateProject (Maybe Natural)

How long, in minutes, from 5 to 480 (8 hours), for CodeBuild to wait before it times out any build that has not been marked as completed. The default is 60 minutes.

createProject_tags :: Lens' CreateProject (Maybe [Tag])

A list of tag key and value pairs associated with this build project.

These tags are available for use by Amazon Web Services services that support CodeBuild build project tags.

createProject_description :: Lens' CreateProject (Maybe Text)

A description that makes the build project easy to identify.

createProject_buildBatchConfig :: Lens' CreateProject (Maybe ProjectBuildBatchConfig)

A ProjectBuildBatchConfig object that defines the batch build options for the project.

createProject_fileSystemLocations :: Lens' CreateProject (Maybe [ProjectFileSystemLocation])

An array of ProjectFileSystemLocation objects for a CodeBuild build project. A ProjectFileSystemLocation object specifies the identifier, location, mountOptions, mountPoint, and type of a file system created using Amazon Elastic File System.

createProject_name :: Lens' CreateProject Text

The name of the build project.

createProject_source :: Lens' CreateProject ProjectSource

Information about the build input source code for the build project.

createProject_artifacts :: Lens' CreateProject ProjectArtifacts

Information about the build output artifacts for the build project.

createProject_environment :: Lens' CreateProject ProjectEnvironment

Information about the build environment for the build project.

createProject_serviceRole :: Lens' CreateProject Text

The ARN of the IAM role that enables CodeBuild to interact with dependent Amazon Web Services services on behalf of the Amazon Web Services account.

Destructuring the Response

data CreateProjectResponse

See: newCreateProjectResponse smart constructor.

Constructors

CreateProjectResponse'

Fields

project :: Maybe Project

Information about the build project that was created.

httpStatus :: Int

The response's http status code.

newCreateProjectResponse

Arguments

:: Int
$sel:httpStatus:CreateProjectResponse'

-> CreateProjectResponse

Create a value of CreateProjectResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:project:CreateProjectResponse', createProjectResponse_project - Information about the build project that was created.

$sel:httpStatus:CreateProjectResponse', createProjectResponse_httpStatus - The response's http status code.

Response Lenses

createProjectResponse_project :: Lens' CreateProjectResponse (Maybe Project)

Information about the build project that was created.

createProjectResponse_httpStatus :: Lens' CreateProjectResponse Int

The response's http status code.