Network.AWS.CodeBuild.CreateReportGroup

Description

Creates a report group. A report group contains a collection of reports.

Creating a Request

data CreateReportGroup

See: newCreateReportGroup smart constructor.

Constructors

CreateReportGroup'

Fields

tags :: Maybe [Tag]

A list of tag key and value pairs associated with this report group.

These tags are available for use by Amazon Web Services services that support CodeBuild report group tags.

name :: Text

The name of the report group.

type' :: ReportType

The type of report group.

exportConfig :: ReportExportConfig

A ReportExportConfig object that contains information about where the report group test results are exported.

newCreateReportGroup

Arguments

:: Text
$sel:name:CreateReportGroup'

-> ReportType
$sel:type':CreateReportGroup'

-> ReportExportConfig
$sel:exportConfig:CreateReportGroup'

-> CreateReportGroup

Create a value of CreateReportGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateReportGroup', createReportGroup_tags - A list of tag key and value pairs associated with this report group.

These tags are available for use by Amazon Web Services services that support CodeBuild report group tags.

$sel:name:CreateReportGroup', createReportGroup_name - The name of the report group.

$sel:type':CreateReportGroup', createReportGroup_type - The type of report group.

$sel:exportConfig:CreateReportGroup', createReportGroup_exportConfig - A ReportExportConfig object that contains information about where the report group test results are exported.

Request Lenses

createReportGroup_tags :: Lens' CreateReportGroup (Maybe [Tag])

A list of tag key and value pairs associated with this report group.

These tags are available for use by Amazon Web Services services that support CodeBuild report group tags.

createReportGroup_name :: Lens' CreateReportGroup Text

The name of the report group.

createReportGroup_type :: Lens' CreateReportGroup ReportType

The type of report group.

createReportGroup_exportConfig :: Lens' CreateReportGroup ReportExportConfig

A ReportExportConfig object that contains information about where the report group test results are exported.

Destructuring the Response

data CreateReportGroupResponse

See: newCreateReportGroupResponse smart constructor.

Constructors

CreateReportGroupResponse'

Fields

reportGroup :: Maybe ReportGroup

Information about the report group that was created.

httpStatus :: Int

The response's http status code.

newCreateReportGroupResponse

Arguments

:: Int
$sel:httpStatus:CreateReportGroupResponse'

-> CreateReportGroupResponse

Create a value of CreateReportGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reportGroup:CreateReportGroupResponse', createReportGroupResponse_reportGroup - Information about the report group that was created.

$sel:httpStatus:CreateReportGroupResponse', createReportGroupResponse_httpStatus - The response's http status code.

Response Lenses

createReportGroupResponse_reportGroup :: Lens' CreateReportGroupResponse (Maybe ReportGroup)

Information about the report group that was created.

createReportGroupResponse_httpStatus :: Lens' CreateReportGroupResponse Int

The response's http status code.