Network.AWS.CodeBuild.DeleteBuildBatch

Description

Deletes a batch build.

Creating a Request

data DeleteBuildBatch

See: newDeleteBuildBatch smart constructor.

Constructors

DeleteBuildBatch'

Fields

id :: Text

The identifier of the batch build to delete.

newDeleteBuildBatch

Arguments

:: Text
$sel:id:DeleteBuildBatch'

-> DeleteBuildBatch

Create a value of DeleteBuildBatch with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:DeleteBuildBatch', deleteBuildBatch_id - The identifier of the batch build to delete.

Request Lenses

deleteBuildBatch_id :: Lens' DeleteBuildBatch Text

The identifier of the batch build to delete.

Destructuring the Response

data DeleteBuildBatchResponse

See: newDeleteBuildBatchResponse smart constructor.

Constructors

DeleteBuildBatchResponse'

Fields

statusCode :: Maybe Text

The status code.

buildsDeleted :: Maybe (NonEmpty Text)

An array of strings that contain the identifiers of the builds that were deleted.

buildsNotDeleted :: Maybe [BuildNotDeleted]

An array of BuildNotDeleted objects that specify the builds that could not be deleted.

httpStatus :: Int

The response's http status code.

newDeleteBuildBatchResponse

Arguments

:: Int
$sel:httpStatus:DeleteBuildBatchResponse'

-> DeleteBuildBatchResponse

Create a value of DeleteBuildBatchResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:statusCode:DeleteBuildBatchResponse', deleteBuildBatchResponse_statusCode - The status code.

$sel:buildsDeleted:DeleteBuildBatchResponse', deleteBuildBatchResponse_buildsDeleted - An array of strings that contain the identifiers of the builds that were deleted.

$sel:buildsNotDeleted:DeleteBuildBatchResponse', deleteBuildBatchResponse_buildsNotDeleted - An array of BuildNotDeleted objects that specify the builds that could not be deleted.

$sel:httpStatus:DeleteBuildBatchResponse', deleteBuildBatchResponse_httpStatus - The response's http status code.

Response Lenses

deleteBuildBatchResponse_statusCode :: Lens' DeleteBuildBatchResponse (Maybe Text)

The status code.

deleteBuildBatchResponse_buildsDeleted :: Lens' DeleteBuildBatchResponse (Maybe (NonEmpty Text))

An array of strings that contain the identifiers of the builds that were deleted.

deleteBuildBatchResponse_buildsNotDeleted :: Lens' DeleteBuildBatchResponse (Maybe [BuildNotDeleted])

An array of BuildNotDeleted objects that specify the builds that could not be deleted.

deleteBuildBatchResponse_httpStatus :: Lens' DeleteBuildBatchResponse Int

The response's http status code.