Network.AWS.CodeBuild.DeleteReport

Description

Deletes a report.

Creating a Request

data DeleteReport

See: newDeleteReport smart constructor.

Constructors

DeleteReport'

Fields

arn :: Text

The ARN of the report to delete.

newDeleteReport

Arguments

:: Text
$sel:arn:DeleteReport'

-> DeleteReport

Create a value of DeleteReport with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:DeleteReport', deleteReport_arn - The ARN of the report to delete.

Request Lenses

deleteReport_arn :: Lens' DeleteReport Text

The ARN of the report to delete.

Destructuring the Response

data DeleteReportResponse

See: newDeleteReportResponse smart constructor.

Constructors

DeleteReportResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteReportResponse

Arguments

:: Int
$sel:httpStatus:DeleteReportResponse'

-> DeleteReportResponse

Create a value of DeleteReportResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteReportResponse', deleteReportResponse_httpStatus - The response's http status code.

Response Lenses

deleteReportResponse_httpStatus :: Lens' DeleteReportResponse Int

The response's http status code.