Network.AWS.CodeBuild.GetReportGroupTrend

Description

Analyzes and accumulates test report values for the specified test reports.

Creating a Request

data GetReportGroupTrend

See: newGetReportGroupTrend smart constructor.

Constructors

GetReportGroupTrend'

Fields

numOfReports :: Maybe Natural

The number of reports to analyze. This operation always retrieves the most recent reports.

If this parameter is omitted, the most recent 100 reports are analyzed.

reportGroupArn :: Text

The ARN of the report group that contains the reports to analyze.

trendField :: ReportGroupTrendFieldType

The test report value to accumulate. This must be one of the following values:

Test reports:

DURATION

Accumulate the test run times for the specified reports.

PASS_RATE

Accumulate the percentage of tests that passed for the specified test reports.

TOTAL

Accumulate the total number of tests for the specified test reports.

Code coverage reports:

BRANCH_COVERAGE

Accumulate the branch coverage percentages for the specified test reports.

BRANCHES_COVERED

Accumulate the branches covered values for the specified test reports.

BRANCHES_MISSED

Accumulate the branches missed values for the specified test reports.

LINE_COVERAGE

Accumulate the line coverage percentages for the specified test reports.

LINES_COVERED

Accumulate the lines covered values for the specified test reports.

LINES_MISSED

Accumulate the lines not covered values for the specified test reports.

newGetReportGroupTrend

Arguments

:: Text
$sel:reportGroupArn:GetReportGroupTrend'

-> ReportGroupTrendFieldType
$sel:trendField:GetReportGroupTrend'

-> GetReportGroupTrend

Create a value of GetReportGroupTrend with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:numOfReports:GetReportGroupTrend', getReportGroupTrend_numOfReports - The number of reports to analyze. This operation always retrieves the most recent reports.

If this parameter is omitted, the most recent 100 reports are analyzed.

$sel:reportGroupArn:GetReportGroupTrend', getReportGroupTrend_reportGroupArn - The ARN of the report group that contains the reports to analyze.

$sel:trendField:GetReportGroupTrend', getReportGroupTrend_trendField - The test report value to accumulate. This must be one of the following values:

Test reports:

DURATION

Accumulate the test run times for the specified reports.

PASS_RATE

Accumulate the percentage of tests that passed for the specified test reports.

TOTAL

Accumulate the total number of tests for the specified test reports.

Code coverage reports:

BRANCH_COVERAGE

Accumulate the branch coverage percentages for the specified test reports.

BRANCHES_COVERED

Accumulate the branches covered values for the specified test reports.

BRANCHES_MISSED

Accumulate the branches missed values for the specified test reports.

LINE_COVERAGE

Accumulate the line coverage percentages for the specified test reports.

LINES_COVERED

Accumulate the lines covered values for the specified test reports.

LINES_MISSED

Accumulate the lines not covered values for the specified test reports.

Request Lenses

getReportGroupTrend_numOfReports :: Lens' GetReportGroupTrend (Maybe Natural)

The number of reports to analyze. This operation always retrieves the most recent reports.

If this parameter is omitted, the most recent 100 reports are analyzed.

getReportGroupTrend_reportGroupArn :: Lens' GetReportGroupTrend Text

The ARN of the report group that contains the reports to analyze.

getReportGroupTrend_trendField :: Lens' GetReportGroupTrend ReportGroupTrendFieldType

The test report value to accumulate. This must be one of the following values:

Test reports:

DURATION

Accumulate the test run times for the specified reports.

PASS_RATE

Accumulate the percentage of tests that passed for the specified test reports.

TOTAL

Accumulate the total number of tests for the specified test reports.

Code coverage reports:

BRANCH_COVERAGE

Accumulate the branch coverage percentages for the specified test reports.

BRANCHES_COVERED

Accumulate the branches covered values for the specified test reports.

BRANCHES_MISSED

Accumulate the branches missed values for the specified test reports.

LINE_COVERAGE

Accumulate the line coverage percentages for the specified test reports.

LINES_COVERED

Accumulate the lines covered values for the specified test reports.

LINES_MISSED

Accumulate the lines not covered values for the specified test reports.

Destructuring the Response

data GetReportGroupTrendResponse

See: newGetReportGroupTrendResponse smart constructor.

Constructors

GetReportGroupTrendResponse'

Fields

rawData :: Maybe [ReportWithRawData]

An array that contains the raw data for each report.

stats :: Maybe ReportGroupTrendStats

Contains the accumulated trend data.

httpStatus :: Int

The response's http status code.

newGetReportGroupTrendResponse

Arguments

:: Int
$sel:httpStatus:GetReportGroupTrendResponse'

-> GetReportGroupTrendResponse

Create a value of GetReportGroupTrendResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:rawData:GetReportGroupTrendResponse', getReportGroupTrendResponse_rawData - An array that contains the raw data for each report.

$sel:stats:GetReportGroupTrendResponse', getReportGroupTrendResponse_stats - Contains the accumulated trend data.

$sel:httpStatus:GetReportGroupTrendResponse', getReportGroupTrendResponse_httpStatus - The response's http status code.

Response Lenses

getReportGroupTrendResponse_rawData :: Lens' GetReportGroupTrendResponse (Maybe [ReportWithRawData])

An array that contains the raw data for each report.

getReportGroupTrendResponse_stats :: Lens' GetReportGroupTrendResponse (Maybe ReportGroupTrendStats)

Contains the accumulated trend data.

getReportGroupTrendResponse_httpStatus :: Lens' GetReportGroupTrendResponse Int

The response's http status code.