Network.AWS.CodeBuild.InvalidateProjectCache

Description

Resets the cache for a project.

Creating a Request

data InvalidateProjectCache

See: newInvalidateProjectCache smart constructor.

Constructors

InvalidateProjectCache'

Fields

projectName :: Text

The name of the CodeBuild build project that the cache is reset for.

newInvalidateProjectCache

Arguments

:: Text
$sel:projectName:InvalidateProjectCache'

-> InvalidateProjectCache

Create a value of InvalidateProjectCache with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:projectName:InvalidateProjectCache', invalidateProjectCache_projectName - The name of the CodeBuild build project that the cache is reset for.

Request Lenses

invalidateProjectCache_projectName :: Lens' InvalidateProjectCache Text

The name of the CodeBuild build project that the cache is reset for.

Destructuring the Response

data InvalidateProjectCacheResponse

See: newInvalidateProjectCacheResponse smart constructor.

Constructors

InvalidateProjectCacheResponse'

Fields

httpStatus :: Int

The response's http status code.

newInvalidateProjectCacheResponse

Arguments

:: Int
$sel:httpStatus:InvalidateProjectCacheResponse'

-> InvalidateProjectCacheResponse

Create a value of InvalidateProjectCacheResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:InvalidateProjectCacheResponse', invalidateProjectCacheResponse_httpStatus - The response's http status code.

Response Lenses

invalidateProjectCacheResponse_httpStatus :: Lens' InvalidateProjectCacheResponse Int

The response's http status code.