Network.AWS.CodeBuild.ListBuildBatches

Description

Retrieves the identifiers of your build batches in the current region.

This operation returns paginated results.

Creating a Request

data ListBuildBatches

See: newListBuildBatches smart constructor.

Constructors

ListBuildBatches'

Fields

nextToken :: Maybe Text

The nextToken value returned from a previous call to ListBuildBatches. This specifies the next item to return. To return the beginning of the list, exclude this parameter.

sortOrder :: Maybe SortOrderType

Specifies the sort order of the returned items. Valid values include:

ASCENDING: List the batch build identifiers in ascending order by identifier.

DESCENDING: List the batch build identifiers in descending order by identifier.

maxResults :: Maybe Natural

The maximum number of results to return.

filter' :: Maybe BuildBatchFilter

A BuildBatchFilter object that specifies the filters for the search.

newListBuildBatches :: ListBuildBatches

Create a value of ListBuildBatches with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListBuildBatches', listBuildBatches_nextToken - The nextToken value returned from a previous call to ListBuildBatches. This specifies the next item to return. To return the beginning of the list, exclude this parameter.

$sel:sortOrder:ListBuildBatches', listBuildBatches_sortOrder - Specifies the sort order of the returned items. Valid values include:

ASCENDING: List the batch build identifiers in ascending order by identifier.

DESCENDING: List the batch build identifiers in descending order by identifier.

$sel:maxResults:ListBuildBatches', listBuildBatches_maxResults - The maximum number of results to return.

$sel:filter':ListBuildBatches', listBuildBatches_filter - A BuildBatchFilter object that specifies the filters for the search.

Request Lenses

listBuildBatches_nextToken :: Lens' ListBuildBatches (Maybe Text)

The nextToken value returned from a previous call to ListBuildBatches. This specifies the next item to return. To return the beginning of the list, exclude this parameter.

listBuildBatches_sortOrder :: Lens' ListBuildBatches (Maybe SortOrderType)

Specifies the sort order of the returned items. Valid values include:

ASCENDING: List the batch build identifiers in ascending order by identifier.

DESCENDING: List the batch build identifiers in descending order by identifier.

listBuildBatches_maxResults :: Lens' ListBuildBatches (Maybe Natural)

The maximum number of results to return.

listBuildBatches_filter :: Lens' ListBuildBatches (Maybe BuildBatchFilter)

A BuildBatchFilter object that specifies the filters for the search.

Destructuring the Response

data ListBuildBatchesResponse

See: newListBuildBatchesResponse smart constructor.

Constructors

ListBuildBatchesResponse'

Fields

nextToken :: Maybe Text

If there are more items to return, this contains a token that is passed to a subsequent call to ListBuildBatches to retrieve the next set of items.

ids :: Maybe [Text]

An array of strings that contains the batch build identifiers.

httpStatus :: Int

The response's http status code.

newListBuildBatchesResponse

Arguments

:: Int
$sel:httpStatus:ListBuildBatchesResponse'

-> ListBuildBatchesResponse

Create a value of ListBuildBatchesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListBuildBatches', listBuildBatchesResponse_nextToken - If there are more items to return, this contains a token that is passed to a subsequent call to ListBuildBatches to retrieve the next set of items.

$sel:ids:ListBuildBatchesResponse', listBuildBatchesResponse_ids - An array of strings that contains the batch build identifiers.

$sel:httpStatus:ListBuildBatchesResponse', listBuildBatchesResponse_httpStatus - The response's http status code.

Response Lenses

listBuildBatchesResponse_nextToken :: Lens' ListBuildBatchesResponse (Maybe Text)

If there are more items to return, this contains a token that is passed to a subsequent call to ListBuildBatches to retrieve the next set of items.

listBuildBatchesResponse_ids :: Lens' ListBuildBatchesResponse (Maybe [Text])

An array of strings that contains the batch build identifiers.

listBuildBatchesResponse_httpStatus :: Lens' ListBuildBatchesResponse Int

The response's http status code.