Network.AWS.CodeBuild.ListReportGroups

Description

Gets a list ARNs for the report groups in the current Amazon Web Services account.

This operation returns paginated results.

Creating a Request

data ListReportGroups

See: newListReportGroups smart constructor.

Constructors

ListReportGroups'

Fields

nextToken :: Maybe Text

During a previous call, the maximum number of items that can be returned is the value specified in maxResults. If there more items in the list, then a unique string called a nextToken is returned. To get the next batch of items in the list, call this operation again, adding the next token to the call. To get all of the items in the list, keep calling this operation with each subsequent next token that is returned, until no more next tokens are returned.

sortOrder :: Maybe SortOrderType

Used to specify the order to sort the list of returned report groups. Valid values are ASCENDING and DESCENDING.

maxResults :: Maybe Natural

The maximum number of paginated report groups returned per response. Use nextToken to iterate pages in the list of returned ReportGroup objects. The default value is 100.

sortBy :: Maybe ReportGroupSortByType

The criterion to be used to list build report groups. Valid values include:

CREATED_TIME: List based on when each report group was created.

LAST_MODIFIED_TIME: List based on when each report group was last changed.

NAME: List based on each report group's name.

newListReportGroups :: ListReportGroups

Create a value of ListReportGroups with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListReportGroups', listReportGroups_nextToken - During a previous call, the maximum number of items that can be returned is the value specified in maxResults. If there more items in the list, then a unique string called a nextToken is returned. To get the next batch of items in the list, call this operation again, adding the next token to the call. To get all of the items in the list, keep calling this operation with each subsequent next token that is returned, until no more next tokens are returned.

$sel:sortOrder:ListReportGroups', listReportGroups_sortOrder - Used to specify the order to sort the list of returned report groups. Valid values are ASCENDING and DESCENDING.

$sel:maxResults:ListReportGroups', listReportGroups_maxResults - The maximum number of paginated report groups returned per response. Use nextToken to iterate pages in the list of returned ReportGroup objects. The default value is 100.

$sel:sortBy:ListReportGroups', listReportGroups_sortBy - The criterion to be used to list build report groups. Valid values include:

CREATED_TIME: List based on when each report group was created.

LAST_MODIFIED_TIME: List based on when each report group was last changed.

NAME: List based on each report group's name.

Request Lenses

listReportGroups_nextToken :: Lens' ListReportGroups (Maybe Text)

During a previous call, the maximum number of items that can be returned is the value specified in maxResults. If there more items in the list, then a unique string called a nextToken is returned. To get the next batch of items in the list, call this operation again, adding the next token to the call. To get all of the items in the list, keep calling this operation with each subsequent next token that is returned, until no more next tokens are returned.

listReportGroups_sortOrder :: Lens' ListReportGroups (Maybe SortOrderType)

Used to specify the order to sort the list of returned report groups. Valid values are ASCENDING and DESCENDING.

listReportGroups_maxResults :: Lens' ListReportGroups (Maybe Natural)

The maximum number of paginated report groups returned per response. Use nextToken to iterate pages in the list of returned ReportGroup objects. The default value is 100.

listReportGroups_sortBy :: Lens' ListReportGroups (Maybe ReportGroupSortByType)

The criterion to be used to list build report groups. Valid values include:

CREATED_TIME: List based on when each report group was created.

LAST_MODIFIED_TIME: List based on when each report group was last changed.

NAME: List based on each report group's name.

Destructuring the Response

data ListReportGroupsResponse

See: newListReportGroupsResponse smart constructor.

Constructors

ListReportGroupsResponse'

Fields

nextToken :: Maybe Text

During a previous call, the maximum number of items that can be returned is the value specified in maxResults. If there more items in the list, then a unique string called a nextToken is returned. To get the next batch of items in the list, call this operation again, adding the next token to the call. To get all of the items in the list, keep calling this operation with each subsequent next token that is returned, until no more next tokens are returned.

reportGroups :: Maybe (NonEmpty Text)

The list of ARNs for the report groups in the current Amazon Web Services account.

httpStatus :: Int

The response's http status code.

newListReportGroupsResponse

Arguments

:: Int
$sel:httpStatus:ListReportGroupsResponse'

-> ListReportGroupsResponse

Create a value of ListReportGroupsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListReportGroups', listReportGroupsResponse_nextToken - During a previous call, the maximum number of items that can be returned is the value specified in maxResults. If there more items in the list, then a unique string called a nextToken is returned. To get the next batch of items in the list, call this operation again, adding the next token to the call. To get all of the items in the list, keep calling this operation with each subsequent next token that is returned, until no more next tokens are returned.

$sel:reportGroups:ListReportGroupsResponse', listReportGroupsResponse_reportGroups - The list of ARNs for the report groups in the current Amazon Web Services account.

$sel:httpStatus:ListReportGroupsResponse', listReportGroupsResponse_httpStatus - The response's http status code.

Response Lenses

listReportGroupsResponse_nextToken :: Lens' ListReportGroupsResponse (Maybe Text)

During a previous call, the maximum number of items that can be returned is the value specified in maxResults. If there more items in the list, then a unique string called a nextToken is returned. To get the next batch of items in the list, call this operation again, adding the next token to the call. To get all of the items in the list, keep calling this operation with each subsequent next token that is returned, until no more next tokens are returned.

listReportGroupsResponse_reportGroups :: Lens' ListReportGroupsResponse (Maybe (NonEmpty Text))

The list of ARNs for the report groups in the current Amazon Web Services account.

listReportGroupsResponse_httpStatus :: Lens' ListReportGroupsResponse Int

The response's http status code.