Network.AWS.CodeBuild.ListSharedProjects

Description

Gets a list of projects that are shared with other Amazon Web Services accounts or users.

This operation returns paginated results.

Creating a Request

data ListSharedProjects

See: newListSharedProjects smart constructor.

Constructors

ListSharedProjects'

Fields

nextToken :: Maybe Text

During a previous call, the maximum number of items that can be returned is the value specified in maxResults. If there more items in the list, then a unique string called a nextToken is returned. To get the next batch of items in the list, call this operation again, adding the next token to the call. To get all of the items in the list, keep calling this operation with each subsequent next token that is returned, until no more next tokens are returned.

sortOrder :: Maybe SortOrderType

The order in which to list shared build projects. Valid values include:

ASCENDING: List in ascending order.

DESCENDING: List in descending order.

maxResults :: Maybe Natural

The maximum number of paginated shared build projects returned per response. Use nextToken to iterate pages in the list of returned Project objects. The default value is 100.

sortBy :: Maybe SharedResourceSortByType

The criterion to be used to list build projects shared with the current Amazon Web Services account or user. Valid values include:

ARN: List based on the ARN.

MODIFIED_TIME: List based on when information about the shared project was last changed.

newListSharedProjects :: ListSharedProjects

Create a value of ListSharedProjects with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListSharedProjects', listSharedProjects_nextToken - During a previous call, the maximum number of items that can be returned is the value specified in maxResults. If there more items in the list, then a unique string called a nextToken is returned. To get the next batch of items in the list, call this operation again, adding the next token to the call. To get all of the items in the list, keep calling this operation with each subsequent next token that is returned, until no more next tokens are returned.

$sel:sortOrder:ListSharedProjects', listSharedProjects_sortOrder - The order in which to list shared build projects. Valid values include:

ASCENDING: List in ascending order.

DESCENDING: List in descending order.

$sel:maxResults:ListSharedProjects', listSharedProjects_maxResults - The maximum number of paginated shared build projects returned per response. Use nextToken to iterate pages in the list of returned Project objects. The default value is 100.

$sel:sortBy:ListSharedProjects', listSharedProjects_sortBy - The criterion to be used to list build projects shared with the current Amazon Web Services account or user. Valid values include:

ARN: List based on the ARN.

MODIFIED_TIME: List based on when information about the shared project was last changed.

Request Lenses

listSharedProjects_nextToken :: Lens' ListSharedProjects (Maybe Text)

During a previous call, the maximum number of items that can be returned is the value specified in maxResults. If there more items in the list, then a unique string called a nextToken is returned. To get the next batch of items in the list, call this operation again, adding the next token to the call. To get all of the items in the list, keep calling this operation with each subsequent next token that is returned, until no more next tokens are returned.

listSharedProjects_sortOrder :: Lens' ListSharedProjects (Maybe SortOrderType)

The order in which to list shared build projects. Valid values include:

ASCENDING: List in ascending order.

DESCENDING: List in descending order.

listSharedProjects_maxResults :: Lens' ListSharedProjects (Maybe Natural)

The maximum number of paginated shared build projects returned per response. Use nextToken to iterate pages in the list of returned Project objects. The default value is 100.

listSharedProjects_sortBy :: Lens' ListSharedProjects (Maybe SharedResourceSortByType)

The criterion to be used to list build projects shared with the current Amazon Web Services account or user. Valid values include:

ARN: List based on the ARN.

MODIFIED_TIME: List based on when information about the shared project was last changed.

Destructuring the Response

data ListSharedProjectsResponse

See: newListSharedProjectsResponse smart constructor.

Constructors

ListSharedProjectsResponse'

Fields

nextToken :: Maybe Text

During a previous call, the maximum number of items that can be returned is the value specified in maxResults. If there more items in the list, then a unique string called a nextToken is returned. To get the next batch of items in the list, call this operation again, adding the next token to the call. To get all of the items in the list, keep calling this operation with each subsequent next token that is returned, until no more next tokens are returned.

projects :: Maybe (NonEmpty Text)

The list of ARNs for the build projects shared with the current Amazon Web Services account or user.

httpStatus :: Int

The response's http status code.

newListSharedProjectsResponse

Arguments

:: Int
$sel:httpStatus:ListSharedProjectsResponse'

-> ListSharedProjectsResponse

Create a value of ListSharedProjectsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListSharedProjects', listSharedProjectsResponse_nextToken - During a previous call, the maximum number of items that can be returned is the value specified in maxResults. If there more items in the list, then a unique string called a nextToken is returned. To get the next batch of items in the list, call this operation again, adding the next token to the call. To get all of the items in the list, keep calling this operation with each subsequent next token that is returned, until no more next tokens are returned.

$sel:projects:ListSharedProjectsResponse', listSharedProjectsResponse_projects - The list of ARNs for the build projects shared with the current Amazon Web Services account or user.

$sel:httpStatus:ListSharedProjectsResponse', listSharedProjectsResponse_httpStatus - The response's http status code.

Response Lenses

listSharedProjectsResponse_nextToken :: Lens' ListSharedProjectsResponse (Maybe Text)

During a previous call, the maximum number of items that can be returned is the value specified in maxResults. If there more items in the list, then a unique string called a nextToken is returned. To get the next batch of items in the list, call this operation again, adding the next token to the call. To get all of the items in the list, keep calling this operation with each subsequent next token that is returned, until no more next tokens are returned.

listSharedProjectsResponse_projects :: Lens' ListSharedProjectsResponse (Maybe (NonEmpty Text))

The list of ARNs for the build projects shared with the current Amazon Web Services account or user.

listSharedProjectsResponse_httpStatus :: Lens' ListSharedProjectsResponse Int

The response's http status code.