Network.AWS.CodeBuild.ListSourceCredentials

Description

Returns a list of SourceCredentialsInfo objects.

Creating a Request

data ListSourceCredentials

See: newListSourceCredentials smart constructor.

Constructors

ListSourceCredentials'

newListSourceCredentials :: ListSourceCredentials

Create a value of ListSourceCredentials with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

Destructuring the Response

data ListSourceCredentialsResponse

See: newListSourceCredentialsResponse smart constructor.

Constructors

ListSourceCredentialsResponse'

Fields

sourceCredentialsInfos :: Maybe [SourceCredentialsInfo]

A list of SourceCredentialsInfo objects. Each SourceCredentialsInfo object includes the authentication type, token ARN, and type of source provider for one set of credentials.

httpStatus :: Int

The response's http status code.

newListSourceCredentialsResponse

Arguments

:: Int
$sel:httpStatus:ListSourceCredentialsResponse'

-> ListSourceCredentialsResponse

Create a value of ListSourceCredentialsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceCredentialsInfos:ListSourceCredentialsResponse', listSourceCredentialsResponse_sourceCredentialsInfos - A list of SourceCredentialsInfo objects. Each SourceCredentialsInfo object includes the authentication type, token ARN, and type of source provider for one set of credentials.

$sel:httpStatus:ListSourceCredentialsResponse', listSourceCredentialsResponse_httpStatus - The response's http status code.

Response Lenses

listSourceCredentialsResponse_sourceCredentialsInfos :: Lens' ListSourceCredentialsResponse (Maybe [SourceCredentialsInfo])

A list of SourceCredentialsInfo objects. Each SourceCredentialsInfo object includes the authentication type, token ARN, and type of source provider for one set of credentials.

listSourceCredentialsResponse_httpStatus :: Lens' ListSourceCredentialsResponse Int

The response's http status code.