Network.AWS.CodeBuild.RetryBuild

Description

Restarts a build.

Creating a Request

data RetryBuild

See: newRetryBuild smart constructor.

Constructors

RetryBuild'

Fields

idempotencyToken :: Maybe Text

A unique, case sensitive identifier you provide to ensure the idempotency of the RetryBuild request. The token is included in the RetryBuild request and is valid for five minutes. If you repeat the RetryBuild request with the same token, but change a parameter, CodeBuild returns a parameter mismatch error.

id :: Maybe Text

Specifies the identifier of the build to restart.

newRetryBuild :: RetryBuild

Create a value of RetryBuild with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:idempotencyToken:RetryBuild', retryBuild_idempotencyToken - A unique, case sensitive identifier you provide to ensure the idempotency of the RetryBuild request. The token is included in the RetryBuild request and is valid for five minutes. If you repeat the RetryBuild request with the same token, but change a parameter, CodeBuild returns a parameter mismatch error.

$sel:id:RetryBuild', retryBuild_id - Specifies the identifier of the build to restart.

Request Lenses

retryBuild_idempotencyToken :: Lens' RetryBuild (Maybe Text)

A unique, case sensitive identifier you provide to ensure the idempotency of the RetryBuild request. The token is included in the RetryBuild request and is valid for five minutes. If you repeat the RetryBuild request with the same token, but change a parameter, CodeBuild returns a parameter mismatch error.

retryBuild_id :: Lens' RetryBuild (Maybe Text)

Specifies the identifier of the build to restart.

Destructuring the Response

data RetryBuildResponse

See: newRetryBuildResponse smart constructor.

Constructors

RetryBuildResponse'

Fields

build :: Maybe Build

httpStatus :: Int

The response's http status code.

newRetryBuildResponse

Arguments

:: Int
$sel:httpStatus:RetryBuildResponse'

-> RetryBuildResponse

Create a value of RetryBuildResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:build:RetryBuildResponse', retryBuildResponse_build - Undocumented member.

$sel:httpStatus:RetryBuildResponse', retryBuildResponse_httpStatus - The response's http status code.

Response Lenses

retryBuildResponse_build :: Lens' RetryBuildResponse (Maybe Build)

Undocumented member.

retryBuildResponse_httpStatus :: Lens' RetryBuildResponse Int

The response's http status code.