Network.AWS.CodeBuild.StartBuildBatch

Description

Starts a batch build for a project.

Creating a Request

data StartBuildBatch

See: newStartBuildBatch smart constructor.

Constructors

StartBuildBatch'

Fields

encryptionKeyOverride :: Maybe Text

The Key Management Service customer master key (CMK) that overrides the one specified in the batch build project. The CMK key encrypts the build output artifacts.

You can use a cross-account KMS key to encrypt the build output artifacts if your service role has permission to that key.

You can specify either the Amazon Resource Name (ARN) of the CMK or, if available, the CMK's alias (using the format alias/<alias-name>).

sourceLocationOverride :: Maybe Text

A location that overrides, for this batch build, the source location defined in the batch build project.

buildBatchConfigOverride :: Maybe ProjectBuildBatchConfig

A BuildBatchConfigOverride object that contains batch build configuration overrides.

environmentVariablesOverride :: Maybe [EnvironmentVariable]

An array of EnvironmentVariable objects that override, or add to, the environment variables defined in the batch build project.

idempotencyToken :: Maybe Text

A unique, case sensitive identifier you provide to ensure the idempotency of the StartBuildBatch request. The token is included in the StartBuildBatch request and is valid for five minutes. If you repeat the StartBuildBatch request with the same token, but change a parameter, CodeBuild returns a parameter mismatch error.

debugSessionEnabled :: Maybe Bool

Specifies if session debugging is enabled for this batch build. For more information, see Viewing a running build in Session Manager. Batch session debugging is not supported for matrix batch builds.

registryCredentialOverride :: Maybe RegistryCredential

A RegistryCredential object that overrides credentials for access to a private registry.

serviceRoleOverride :: Maybe Text

The name of a service role for this batch build that overrides the one specified in the batch build project.

cacheOverride :: Maybe ProjectCache

A ProjectCache object that specifies cache overrides.

queuedTimeoutInMinutesOverride :: Maybe Natural

The number of minutes a batch build is allowed to be queued before it times out.

secondarySourcesOverride :: Maybe [ProjectSource]

An array of ProjectSource objects that override the secondary sources defined in the batch build project.

gitCloneDepthOverride :: Maybe Natural

The user-defined depth of history, with a minimum value of 0, that overrides, for this batch build only, any previous depth of history defined in the batch build project.

imagePullCredentialsTypeOverride :: Maybe ImagePullCredentialsType

The type of credentials CodeBuild uses to pull images in your batch build. There are two valid values:

CODEBUILD

Specifies that CodeBuild uses its own credentials. This requires that you modify your ECR repository policy to trust CodeBuild's service principal.

SERVICE_ROLE

Specifies that CodeBuild uses your build project's service role.

When using a cross-account or private registry image, you must use SERVICE_ROLE credentials. When using an CodeBuild curated image, you must use CODEBUILD credentials.

logsConfigOverride :: Maybe LogsConfig

A LogsConfig object that override the log settings defined in the batch build project.

sourceAuthOverride :: Maybe SourceAuth

A SourceAuth object that overrides the one defined in the batch build project. This override applies only if the build project's source is BitBucket or GitHub.

gitSubmodulesConfigOverride :: Maybe GitSubmodulesConfig

A GitSubmodulesConfig object that overrides the Git submodules configuration for this batch build.

environmentTypeOverride :: Maybe EnvironmentType

A container type for this batch build that overrides the one specified in the batch build project.

certificateOverride :: Maybe Text

The name of a certificate for this batch build that overrides the one specified in the batch build project.

computeTypeOverride :: Maybe ComputeType

The name of a compute type for this batch build that overrides the one specified in the batch build project.

reportBuildBatchStatusOverride :: Maybe Bool

Set to true to report to your source provider the status of a batch build's start and completion. If you use this option with a source provider other than GitHub, GitHub Enterprise, or Bitbucket, an invalidInputException is thrown.

The status of a build triggered by a webhook is always reported to your source provider.

privilegedModeOverride :: Maybe Bool

Enable this flag to override privileged mode in the batch build project.

sourceVersion :: Maybe Text

The version of the batch build input to be built, for this build only. If not specified, the latest version is used. If specified, the contents depends on the source provider:

CodeCommit

The commit ID, branch, or Git tag to use.

GitHub

The commit ID, pull request ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a pull request ID is specified, it must use the format pr/pull-request-ID (for example pr/25). If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

Bitbucket

The commit ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

Amazon S3

The version ID of the object that represents the build input ZIP file to use.

If sourceVersion is specified at the project level, then this sourceVersion (at the build level) takes precedence.

For more information, see Source Version Sample with CodeBuild in the CodeBuild User Guide.

buildspecOverride :: Maybe Text

A buildspec file declaration that overrides, for this build only, the latest one already defined in the build project.

If this value is set, it can be either an inline buildspec definition, the path to an alternate buildspec file relative to the value of the built-in CODEBUILD_SRC_DIR environment variable, or the path to an S3 bucket. The bucket must be in the same Amazon Web Services Region as the build project. Specify the buildspec file using its ARN (for example, arn:aws:s3:::my-codebuild-sample2/buildspec.yml). If this value is not provided or is set to an empty string, the source code must contain a buildspec file in its root directory. For more information, see Buildspec File Name and Storage Location.

secondarySourcesVersionOverride :: Maybe [ProjectSourceVersion]

An array of ProjectSourceVersion objects that override the secondary source versions in the batch build project.

insecureSslOverride :: Maybe Bool

Enable this flag to override the insecure SSL setting that is specified in the batch build project. The insecure SSL setting determines whether to ignore SSL warnings while connecting to the project source code. This override applies only if the build's source is GitHub Enterprise.

imageOverride :: Maybe Text

The name of an image for this batch build that overrides the one specified in the batch build project.

secondaryArtifactsOverride :: Maybe [ProjectArtifacts]

An array of ProjectArtifacts objects that override the secondary artifacts defined in the batch build project.

buildTimeoutInMinutesOverride :: Maybe Natural

Overrides the build timeout specified in the batch build project.

artifactsOverride :: Maybe ProjectArtifacts

An array of ProjectArtifacts objects that contains information about the build output artifact overrides for the build project.

sourceTypeOverride :: Maybe SourceType

The source input type that overrides the source input defined in the batch build project.

projectName :: Text

The name of the project.

newStartBuildBatch

Arguments

:: Text
$sel:projectName:StartBuildBatch'

-> StartBuildBatch

Create a value of StartBuildBatch with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:encryptionKeyOverride:StartBuildBatch', startBuildBatch_encryptionKeyOverride - The Key Management Service customer master key (CMK) that overrides the one specified in the batch build project. The CMK key encrypts the build output artifacts.

You can use a cross-account KMS key to encrypt the build output artifacts if your service role has permission to that key.

You can specify either the Amazon Resource Name (ARN) of the CMK or, if available, the CMK's alias (using the format alias/<alias-name>).

$sel:sourceLocationOverride:StartBuildBatch', startBuildBatch_sourceLocationOverride - A location that overrides, for this batch build, the source location defined in the batch build project.

$sel:buildBatchConfigOverride:StartBuildBatch', startBuildBatch_buildBatchConfigOverride - A BuildBatchConfigOverride object that contains batch build configuration overrides.

$sel:environmentVariablesOverride:StartBuildBatch', startBuildBatch_environmentVariablesOverride - An array of EnvironmentVariable objects that override, or add to, the environment variables defined in the batch build project.

$sel:idempotencyToken:StartBuildBatch', startBuildBatch_idempotencyToken - A unique, case sensitive identifier you provide to ensure the idempotency of the StartBuildBatch request. The token is included in the StartBuildBatch request and is valid for five minutes. If you repeat the StartBuildBatch request with the same token, but change a parameter, CodeBuild returns a parameter mismatch error.

$sel:debugSessionEnabled:StartBuildBatch', startBuildBatch_debugSessionEnabled - Specifies if session debugging is enabled for this batch build. For more information, see Viewing a running build in Session Manager. Batch session debugging is not supported for matrix batch builds.

$sel:registryCredentialOverride:StartBuildBatch', startBuildBatch_registryCredentialOverride - A RegistryCredential object that overrides credentials for access to a private registry.

$sel:serviceRoleOverride:StartBuildBatch', startBuildBatch_serviceRoleOverride - The name of a service role for this batch build that overrides the one specified in the batch build project.

$sel:cacheOverride:StartBuildBatch', startBuildBatch_cacheOverride - A ProjectCache object that specifies cache overrides.

$sel:queuedTimeoutInMinutesOverride:StartBuildBatch', startBuildBatch_queuedTimeoutInMinutesOverride - The number of minutes a batch build is allowed to be queued before it times out.

$sel:secondarySourcesOverride:StartBuildBatch', startBuildBatch_secondarySourcesOverride - An array of ProjectSource objects that override the secondary sources defined in the batch build project.

$sel:gitCloneDepthOverride:StartBuildBatch', startBuildBatch_gitCloneDepthOverride - The user-defined depth of history, with a minimum value of 0, that overrides, for this batch build only, any previous depth of history defined in the batch build project.

$sel:imagePullCredentialsTypeOverride:StartBuildBatch', startBuildBatch_imagePullCredentialsTypeOverride - The type of credentials CodeBuild uses to pull images in your batch build. There are two valid values:

CODEBUILD

Specifies that CodeBuild uses its own credentials. This requires that you modify your ECR repository policy to trust CodeBuild's service principal.

SERVICE_ROLE

Specifies that CodeBuild uses your build project's service role.

When using a cross-account or private registry image, you must use SERVICE_ROLE credentials. When using an CodeBuild curated image, you must use CODEBUILD credentials.

$sel:logsConfigOverride:StartBuildBatch', startBuildBatch_logsConfigOverride - A LogsConfig object that override the log settings defined in the batch build project.

$sel:sourceAuthOverride:StartBuildBatch', startBuildBatch_sourceAuthOverride - A SourceAuth object that overrides the one defined in the batch build project. This override applies only if the build project's source is BitBucket or GitHub.

$sel:gitSubmodulesConfigOverride:StartBuildBatch', startBuildBatch_gitSubmodulesConfigOverride - A GitSubmodulesConfig object that overrides the Git submodules configuration for this batch build.

$sel:environmentTypeOverride:StartBuildBatch', startBuildBatch_environmentTypeOverride - A container type for this batch build that overrides the one specified in the batch build project.

$sel:certificateOverride:StartBuildBatch', startBuildBatch_certificateOverride - The name of a certificate for this batch build that overrides the one specified in the batch build project.

$sel:computeTypeOverride:StartBuildBatch', startBuildBatch_computeTypeOverride - The name of a compute type for this batch build that overrides the one specified in the batch build project.

$sel:reportBuildBatchStatusOverride:StartBuildBatch', startBuildBatch_reportBuildBatchStatusOverride - Set to true to report to your source provider the status of a batch build's start and completion. If you use this option with a source provider other than GitHub, GitHub Enterprise, or Bitbucket, an invalidInputException is thrown.

The status of a build triggered by a webhook is always reported to your source provider.

$sel:privilegedModeOverride:StartBuildBatch', startBuildBatch_privilegedModeOverride - Enable this flag to override privileged mode in the batch build project.

$sel:sourceVersion:StartBuildBatch', startBuildBatch_sourceVersion - The version of the batch build input to be built, for this build only. If not specified, the latest version is used. If specified, the contents depends on the source provider:

CodeCommit

The commit ID, branch, or Git tag to use.

GitHub

The commit ID, pull request ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a pull request ID is specified, it must use the format pr/pull-request-ID (for example pr/25). If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

Bitbucket

The commit ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

Amazon S3

The version ID of the object that represents the build input ZIP file to use.

If sourceVersion is specified at the project level, then this sourceVersion (at the build level) takes precedence.

For more information, see Source Version Sample with CodeBuild in the CodeBuild User Guide.

$sel:buildspecOverride:StartBuildBatch', startBuildBatch_buildspecOverride - A buildspec file declaration that overrides, for this build only, the latest one already defined in the build project.

If this value is set, it can be either an inline buildspec definition, the path to an alternate buildspec file relative to the value of the built-in CODEBUILD_SRC_DIR environment variable, or the path to an S3 bucket. The bucket must be in the same Amazon Web Services Region as the build project. Specify the buildspec file using its ARN (for example, arn:aws:s3:::my-codebuild-sample2/buildspec.yml). If this value is not provided or is set to an empty string, the source code must contain a buildspec file in its root directory. For more information, see Buildspec File Name and Storage Location.

$sel:secondarySourcesVersionOverride:StartBuildBatch', startBuildBatch_secondarySourcesVersionOverride - An array of ProjectSourceVersion objects that override the secondary source versions in the batch build project.

$sel:insecureSslOverride:StartBuildBatch', startBuildBatch_insecureSslOverride - Enable this flag to override the insecure SSL setting that is specified in the batch build project. The insecure SSL setting determines whether to ignore SSL warnings while connecting to the project source code. This override applies only if the build's source is GitHub Enterprise.

$sel:imageOverride:StartBuildBatch', startBuildBatch_imageOverride - The name of an image for this batch build that overrides the one specified in the batch build project.

$sel:secondaryArtifactsOverride:StartBuildBatch', startBuildBatch_secondaryArtifactsOverride - An array of ProjectArtifacts objects that override the secondary artifacts defined in the batch build project.

$sel:buildTimeoutInMinutesOverride:StartBuildBatch', startBuildBatch_buildTimeoutInMinutesOverride - Overrides the build timeout specified in the batch build project.

$sel:artifactsOverride:StartBuildBatch', startBuildBatch_artifactsOverride - An array of ProjectArtifacts objects that contains information about the build output artifact overrides for the build project.

$sel:sourceTypeOverride:StartBuildBatch', startBuildBatch_sourceTypeOverride - The source input type that overrides the source input defined in the batch build project.

$sel:projectName:StartBuildBatch', startBuildBatch_projectName - The name of the project.

Request Lenses

startBuildBatch_encryptionKeyOverride :: Lens' StartBuildBatch (Maybe Text)

The Key Management Service customer master key (CMK) that overrides the one specified in the batch build project. The CMK key encrypts the build output artifacts.

You can use a cross-account KMS key to encrypt the build output artifacts if your service role has permission to that key.

You can specify either the Amazon Resource Name (ARN) of the CMK or, if available, the CMK's alias (using the format alias/<alias-name>).

startBuildBatch_sourceLocationOverride :: Lens' StartBuildBatch (Maybe Text)

A location that overrides, for this batch build, the source location defined in the batch build project.

startBuildBatch_buildBatchConfigOverride :: Lens' StartBuildBatch (Maybe ProjectBuildBatchConfig)

A BuildBatchConfigOverride object that contains batch build configuration overrides.

startBuildBatch_environmentVariablesOverride :: Lens' StartBuildBatch (Maybe [EnvironmentVariable])

An array of EnvironmentVariable objects that override, or add to, the environment variables defined in the batch build project.

startBuildBatch_idempotencyToken :: Lens' StartBuildBatch (Maybe Text)

A unique, case sensitive identifier you provide to ensure the idempotency of the StartBuildBatch request. The token is included in the StartBuildBatch request and is valid for five minutes. If you repeat the StartBuildBatch request with the same token, but change a parameter, CodeBuild returns a parameter mismatch error.

startBuildBatch_debugSessionEnabled :: Lens' StartBuildBatch (Maybe Bool)

Specifies if session debugging is enabled for this batch build. For more information, see Viewing a running build in Session Manager. Batch session debugging is not supported for matrix batch builds.

startBuildBatch_registryCredentialOverride :: Lens' StartBuildBatch (Maybe RegistryCredential)

A RegistryCredential object that overrides credentials for access to a private registry.

startBuildBatch_serviceRoleOverride :: Lens' StartBuildBatch (Maybe Text)

The name of a service role for this batch build that overrides the one specified in the batch build project.

startBuildBatch_cacheOverride :: Lens' StartBuildBatch (Maybe ProjectCache)

A ProjectCache object that specifies cache overrides.

startBuildBatch_queuedTimeoutInMinutesOverride :: Lens' StartBuildBatch (Maybe Natural)

The number of minutes a batch build is allowed to be queued before it times out.

startBuildBatch_secondarySourcesOverride :: Lens' StartBuildBatch (Maybe [ProjectSource])

An array of ProjectSource objects that override the secondary sources defined in the batch build project.

startBuildBatch_gitCloneDepthOverride :: Lens' StartBuildBatch (Maybe Natural)

The user-defined depth of history, with a minimum value of 0, that overrides, for this batch build only, any previous depth of history defined in the batch build project.

startBuildBatch_imagePullCredentialsTypeOverride :: Lens' StartBuildBatch (Maybe ImagePullCredentialsType)

The type of credentials CodeBuild uses to pull images in your batch build. There are two valid values:

CODEBUILD

Specifies that CodeBuild uses its own credentials. This requires that you modify your ECR repository policy to trust CodeBuild's service principal.

SERVICE_ROLE

Specifies that CodeBuild uses your build project's service role.

When using a cross-account or private registry image, you must use SERVICE_ROLE credentials. When using an CodeBuild curated image, you must use CODEBUILD credentials.

startBuildBatch_logsConfigOverride :: Lens' StartBuildBatch (Maybe LogsConfig)

A LogsConfig object that override the log settings defined in the batch build project.

startBuildBatch_sourceAuthOverride :: Lens' StartBuildBatch (Maybe SourceAuth)

A SourceAuth object that overrides the one defined in the batch build project. This override applies only if the build project's source is BitBucket or GitHub.

startBuildBatch_gitSubmodulesConfigOverride :: Lens' StartBuildBatch (Maybe GitSubmodulesConfig)

A GitSubmodulesConfig object that overrides the Git submodules configuration for this batch build.

startBuildBatch_environmentTypeOverride :: Lens' StartBuildBatch (Maybe EnvironmentType)

A container type for this batch build that overrides the one specified in the batch build project.

startBuildBatch_certificateOverride :: Lens' StartBuildBatch (Maybe Text)

The name of a certificate for this batch build that overrides the one specified in the batch build project.

startBuildBatch_computeTypeOverride :: Lens' StartBuildBatch (Maybe ComputeType)

The name of a compute type for this batch build that overrides the one specified in the batch build project.

startBuildBatch_reportBuildBatchStatusOverride :: Lens' StartBuildBatch (Maybe Bool)

Set to true to report to your source provider the status of a batch build's start and completion. If you use this option with a source provider other than GitHub, GitHub Enterprise, or Bitbucket, an invalidInputException is thrown.

The status of a build triggered by a webhook is always reported to your source provider.

startBuildBatch_privilegedModeOverride :: Lens' StartBuildBatch (Maybe Bool)

Enable this flag to override privileged mode in the batch build project.

startBuildBatch_sourceVersion :: Lens' StartBuildBatch (Maybe Text)

The version of the batch build input to be built, for this build only. If not specified, the latest version is used. If specified, the contents depends on the source provider:

CodeCommit

The commit ID, branch, or Git tag to use.

GitHub

The commit ID, pull request ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a pull request ID is specified, it must use the format pr/pull-request-ID (for example pr/25). If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

Bitbucket

The commit ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

Amazon S3

The version ID of the object that represents the build input ZIP file to use.

If sourceVersion is specified at the project level, then this sourceVersion (at the build level) takes precedence.

For more information, see Source Version Sample with CodeBuild in the CodeBuild User Guide.

startBuildBatch_buildspecOverride :: Lens' StartBuildBatch (Maybe Text)

A buildspec file declaration that overrides, for this build only, the latest one already defined in the build project.

If this value is set, it can be either an inline buildspec definition, the path to an alternate buildspec file relative to the value of the built-in CODEBUILD_SRC_DIR environment variable, or the path to an S3 bucket. The bucket must be in the same Amazon Web Services Region as the build project. Specify the buildspec file using its ARN (for example, arn:aws:s3:::my-codebuild-sample2/buildspec.yml). If this value is not provided or is set to an empty string, the source code must contain a buildspec file in its root directory. For more information, see Buildspec File Name and Storage Location.

startBuildBatch_secondarySourcesVersionOverride :: Lens' StartBuildBatch (Maybe [ProjectSourceVersion])

An array of ProjectSourceVersion objects that override the secondary source versions in the batch build project.

startBuildBatch_insecureSslOverride :: Lens' StartBuildBatch (Maybe Bool)

Enable this flag to override the insecure SSL setting that is specified in the batch build project. The insecure SSL setting determines whether to ignore SSL warnings while connecting to the project source code. This override applies only if the build's source is GitHub Enterprise.

startBuildBatch_imageOverride :: Lens' StartBuildBatch (Maybe Text)

The name of an image for this batch build that overrides the one specified in the batch build project.

startBuildBatch_secondaryArtifactsOverride :: Lens' StartBuildBatch (Maybe [ProjectArtifacts])

An array of ProjectArtifacts objects that override the secondary artifacts defined in the batch build project.

startBuildBatch_buildTimeoutInMinutesOverride :: Lens' StartBuildBatch (Maybe Natural)

Overrides the build timeout specified in the batch build project.

startBuildBatch_artifactsOverride :: Lens' StartBuildBatch (Maybe ProjectArtifacts)

An array of ProjectArtifacts objects that contains information about the build output artifact overrides for the build project.

startBuildBatch_sourceTypeOverride :: Lens' StartBuildBatch (Maybe SourceType)

The source input type that overrides the source input defined in the batch build project.

startBuildBatch_projectName :: Lens' StartBuildBatch Text

The name of the project.

Destructuring the Response

data StartBuildBatchResponse

See: newStartBuildBatchResponse smart constructor.

Constructors

StartBuildBatchResponse'

Fields

buildBatch :: Maybe BuildBatch

A BuildBatch object that contains information about the batch build.

httpStatus :: Int

The response's http status code.

newStartBuildBatchResponse

Arguments

:: Int
$sel:httpStatus:StartBuildBatchResponse'

-> StartBuildBatchResponse

Create a value of StartBuildBatchResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:buildBatch:StartBuildBatchResponse', startBuildBatchResponse_buildBatch - A BuildBatch object that contains information about the batch build.

$sel:httpStatus:StartBuildBatchResponse', startBuildBatchResponse_httpStatus - The response's http status code.

Response Lenses

startBuildBatchResponse_buildBatch :: Lens' StartBuildBatchResponse (Maybe BuildBatch)

A BuildBatch object that contains information about the batch build.

startBuildBatchResponse_httpStatus :: Lens' StartBuildBatchResponse Int

The response's http status code.