Network.AWS.CodeBuild.Types.Build

Description

Documentation

data Build

Information about a build.

See: newBuild smart constructor.

Constructors

Build'

Fields

vpcConfig :: Maybe VpcConfig

If your CodeBuild project accesses resources in an Amazon VPC, you provide this parameter that identifies the VPC ID and the list of security group IDs and subnet IDs. The security groups and subnets must belong to the same VPC. You must provide at least one security group and one subnet ID.

sourceVersion :: Maybe Text

Any version identifier for the version of the source code to be built. If sourceVersion is specified at the project level, then this sourceVersion (at the build level) takes precedence.

For more information, see Source Version Sample with CodeBuild in the CodeBuild User Guide.

buildBatchArn :: Maybe Text

The ARN of the batch build that this build is a member of, if applicable.

secondaryArtifacts :: Maybe [BuildArtifacts]

An array of ProjectArtifacts objects.

resolvedSourceVersion :: Maybe Text

An identifier for the version of this build's source code.

For CodeCommit, GitHub, GitHub Enterprise, and BitBucket, the commit ID.

For CodePipeline, the source revision provided by CodePipeline.

For Amazon S3, this does not apply.

phases :: Maybe [BuildPhase]

Information about all previous build phases that are complete and information about any current build phase that is not yet complete.

cache :: Maybe ProjectCache

Information about the cache for the build.

serviceRole :: Maybe Text

The name of a service role used for this build.

secondarySourceVersions :: Maybe [ProjectSourceVersion]

An array of ProjectSourceVersion objects. Each ProjectSourceVersion must be one of:

For CodeCommit: the commit ID, branch, or Git tag to use.

For GitHub: the commit ID, pull request ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a pull request ID is specified, it must use the format pr/pull-request-ID (for example, pr/25). If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

For Bitbucket: the commit ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

For Amazon S3: the version ID of the object that represents the build input ZIP file to use.

networkInterface :: Maybe NetworkInterface

Describes a network interface.

encryptionKey :: Maybe Text

The Key Management Service customer master key (CMK) to be used for encrypting the build output artifacts.

You can use a cross-account KMS key to encrypt the build output artifacts if your service role has permission to that key.

You can specify either the Amazon Resource Name (ARN) of the CMK or, if available, the CMK's alias (using the format alias/<alias-name>).

id :: Maybe Text

The unique ID for the build.

environment :: Maybe ProjectEnvironment

Information about the build environment for this build.

source :: Maybe ProjectSource

Information about the source code to be built.

startTime :: Maybe POSIX

When the build process started, expressed in Unix time format.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the build.

buildNumber :: Maybe Integer

The number of the build. For each project, the buildNumber of its first build is 1. The buildNumber of each subsequent build is incremented by 1. If a build is deleted, the buildNumber of other builds does not change.

artifacts :: Maybe BuildArtifacts

Information about the output artifacts for the build.

projectName :: Maybe Text

The name of the CodeBuild project.

buildStatus :: Maybe StatusType

The current status of the build. Valid values include:

FAILED: The build failed.

FAULT: The build faulted.

IN_PROGRESS: The build is still in progress.

STOPPED: The build stopped.

SUCCEEDED: The build succeeded.

TIMED_OUT: The build timed out.

endTime :: Maybe POSIX

When the build process ended, expressed in Unix time format.

buildComplete :: Maybe Bool

Whether the build is complete. True if complete; otherwise, false.

logs :: Maybe LogsLocation

Information about the build's logs in CloudWatch Logs.

debugSession :: Maybe DebugSession

Contains information about the debug session for this build.

queuedTimeoutInMinutes :: Maybe Int

The number of minutes a build is allowed to be queued before it times out.

secondarySources :: Maybe [ProjectSource]

An array of ProjectSource objects.

timeoutInMinutes :: Maybe Int

How long, in minutes, for CodeBuild to wait before timing out this build if it does not get marked as completed.

currentPhase :: Maybe Text

The current build phase.

initiator :: Maybe Text

The entity that started the build. Valid values include:

If CodePipeline started the build, the pipeline's name (for example, codepipeline/my-demo-pipeline).

If an IAM user started the build, the user's name (for example, MyUserName).

If the Jenkins plugin for CodeBuild started the build, the string CodeBuild-Jenkins-Plugin.

reportArns :: Maybe [Text]

An array of the ARNs associated with this build's reports.

exportedEnvironmentVariables :: Maybe [ExportedEnvironmentVariable]

A list of exported environment variables for this build.

Exported environment variables are used in conjunction with CodePipeline to export environment variables from the current build stage to subsequent stages in the pipeline. For more information, see Working with variables in the CodePipeline User Guide.

fileSystemLocations :: Maybe [ProjectFileSystemLocation]

An array of ProjectFileSystemLocation objects for a CodeBuild build project. A ProjectFileSystemLocation object specifies the identifier, location, mountOptions, mountPoint, and type of a file system created using Amazon Elastic File System.

newBuild :: Build

Create a value of Build with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpcConfig:Build', build_vpcConfig - If your CodeBuild project accesses resources in an Amazon VPC, you provide this parameter that identifies the VPC ID and the list of security group IDs and subnet IDs. The security groups and subnets must belong to the same VPC. You must provide at least one security group and one subnet ID.

$sel:sourceVersion:Build', build_sourceVersion - Any version identifier for the version of the source code to be built. If sourceVersion is specified at the project level, then this sourceVersion (at the build level) takes precedence.

For more information, see Source Version Sample with CodeBuild in the CodeBuild User Guide.

$sel:buildBatchArn:Build', build_buildBatchArn - The ARN of the batch build that this build is a member of, if applicable.

$sel:secondaryArtifacts:Build', build_secondaryArtifacts - An array of ProjectArtifacts objects.

$sel:resolvedSourceVersion:Build', build_resolvedSourceVersion - An identifier for the version of this build's source code.

For CodeCommit, GitHub, GitHub Enterprise, and BitBucket, the commit ID.

For CodePipeline, the source revision provided by CodePipeline.

For Amazon S3, this does not apply.

$sel:phases:Build', build_phases - Information about all previous build phases that are complete and information about any current build phase that is not yet complete.

$sel:cache:Build', build_cache - Information about the cache for the build.

$sel:serviceRole:Build', build_serviceRole - The name of a service role used for this build.

$sel:secondarySourceVersions:Build', build_secondarySourceVersions - An array of ProjectSourceVersion objects. Each ProjectSourceVersion must be one of:

For CodeCommit: the commit ID, branch, or Git tag to use.

For GitHub: the commit ID, pull request ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a pull request ID is specified, it must use the format pr/pull-request-ID (for example, pr/25). If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

For Bitbucket: the commit ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

For Amazon S3: the version ID of the object that represents the build input ZIP file to use.

$sel:networkInterface:Build', build_networkInterface - Describes a network interface.

$sel:encryptionKey:Build', build_encryptionKey - The Key Management Service customer master key (CMK) to be used for encrypting the build output artifacts.

You can use a cross-account KMS key to encrypt the build output artifacts if your service role has permission to that key.

You can specify either the Amazon Resource Name (ARN) of the CMK or, if available, the CMK's alias (using the format alias/<alias-name>).

$sel:id:Build', build_id - The unique ID for the build.

$sel:environment:Build', build_environment - Information about the build environment for this build.

$sel:source:Build', build_source - Information about the source code to be built.

$sel:startTime:Build', build_startTime - When the build process started, expressed in Unix time format.

$sel:arn:Build', build_arn - The Amazon Resource Name (ARN) of the build.

$sel:buildNumber:Build', build_buildNumber - The number of the build. For each project, the buildNumber of its first build is 1. The buildNumber of each subsequent build is incremented by 1. If a build is deleted, the buildNumber of other builds does not change.

$sel:artifacts:Build', build_artifacts - Information about the output artifacts for the build.

$sel:projectName:Build', build_projectName - The name of the CodeBuild project.

$sel:buildStatus:Build', build_buildStatus - The current status of the build. Valid values include:

FAILED: The build failed.

FAULT: The build faulted.

IN_PROGRESS: The build is still in progress.

STOPPED: The build stopped.

SUCCEEDED: The build succeeded.

TIMED_OUT: The build timed out.

$sel:endTime:Build', build_endTime - When the build process ended, expressed in Unix time format.

$sel:buildComplete:Build', build_buildComplete - Whether the build is complete. True if complete; otherwise, false.

$sel:logs:Build', build_logs - Information about the build's logs in CloudWatch Logs.

$sel:debugSession:Build', build_debugSession - Contains information about the debug session for this build.

$sel:queuedTimeoutInMinutes:Build', build_queuedTimeoutInMinutes - The number of minutes a build is allowed to be queued before it times out.

$sel:secondarySources:Build', build_secondarySources - An array of ProjectSource objects.

$sel:timeoutInMinutes:Build', build_timeoutInMinutes - How long, in minutes, for CodeBuild to wait before timing out this build if it does not get marked as completed.

$sel:currentPhase:Build', build_currentPhase - The current build phase.

$sel:initiator:Build', build_initiator - The entity that started the build. Valid values include:

If CodePipeline started the build, the pipeline's name (for example, codepipeline/my-demo-pipeline).

If an IAM user started the build, the user's name (for example, MyUserName).

If the Jenkins plugin for CodeBuild started the build, the string CodeBuild-Jenkins-Plugin.

$sel:reportArns:Build', build_reportArns - An array of the ARNs associated with this build's reports.

$sel:exportedEnvironmentVariables:Build', build_exportedEnvironmentVariables - A list of exported environment variables for this build.

Exported environment variables are used in conjunction with CodePipeline to export environment variables from the current build stage to subsequent stages in the pipeline. For more information, see Working with variables in the CodePipeline User Guide.

$sel:fileSystemLocations:Build', build_fileSystemLocations - An array of ProjectFileSystemLocation objects for a CodeBuild build project. A ProjectFileSystemLocation object specifies the identifier, location, mountOptions, mountPoint, and type of a file system created using Amazon Elastic File System.

build_vpcConfig :: Lens' Build (Maybe VpcConfig)

If your CodeBuild project accesses resources in an Amazon VPC, you provide this parameter that identifies the VPC ID and the list of security group IDs and subnet IDs. The security groups and subnets must belong to the same VPC. You must provide at least one security group and one subnet ID.

build_sourceVersion :: Lens' Build (Maybe Text)

Any version identifier for the version of the source code to be built. If sourceVersion is specified at the project level, then this sourceVersion (at the build level) takes precedence.

For more information, see Source Version Sample with CodeBuild in the CodeBuild User Guide.

build_buildBatchArn :: Lens' Build (Maybe Text)

The ARN of the batch build that this build is a member of, if applicable.

build_secondaryArtifacts :: Lens' Build (Maybe [BuildArtifacts])

An array of ProjectArtifacts objects.

build_resolvedSourceVersion :: Lens' Build (Maybe Text)

An identifier for the version of this build's source code.

For CodeCommit, GitHub, GitHub Enterprise, and BitBucket, the commit ID.

For CodePipeline, the source revision provided by CodePipeline.

For Amazon S3, this does not apply.

build_phases :: Lens' Build (Maybe [BuildPhase])

Information about all previous build phases that are complete and information about any current build phase that is not yet complete.

build_cache :: Lens' Build (Maybe ProjectCache)

Information about the cache for the build.

build_serviceRole :: Lens' Build (Maybe Text)

The name of a service role used for this build.

build_secondarySourceVersions :: Lens' Build (Maybe [ProjectSourceVersion])

An array of ProjectSourceVersion objects. Each ProjectSourceVersion must be one of:

For CodeCommit: the commit ID, branch, or Git tag to use.

For GitHub: the commit ID, pull request ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a pull request ID is specified, it must use the format pr/pull-request-ID (for example, pr/25). If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

For Bitbucket: the commit ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

For Amazon S3: the version ID of the object that represents the build input ZIP file to use.

build_networkInterface :: Lens' Build (Maybe NetworkInterface)

Describes a network interface.

build_encryptionKey :: Lens' Build (Maybe Text)

The Key Management Service customer master key (CMK) to be used for encrypting the build output artifacts.

You can use a cross-account KMS key to encrypt the build output artifacts if your service role has permission to that key.

You can specify either the Amazon Resource Name (ARN) of the CMK or, if available, the CMK's alias (using the format alias/<alias-name>).

build_id :: Lens' Build (Maybe Text)

The unique ID for the build.

build_environment :: Lens' Build (Maybe ProjectEnvironment)

Information about the build environment for this build.

build_source :: Lens' Build (Maybe ProjectSource)

Information about the source code to be built.

build_startTime :: Lens' Build (Maybe UTCTime)

When the build process started, expressed in Unix time format.

build_arn :: Lens' Build (Maybe Text)

The Amazon Resource Name (ARN) of the build.

build_buildNumber :: Lens' Build (Maybe Integer)

The number of the build. For each project, the buildNumber of its first build is 1. The buildNumber of each subsequent build is incremented by 1. If a build is deleted, the buildNumber of other builds does not change.

build_artifacts :: Lens' Build (Maybe BuildArtifacts)

Information about the output artifacts for the build.

build_projectName :: Lens' Build (Maybe Text)

The name of the CodeBuild project.

build_buildStatus :: Lens' Build (Maybe StatusType)

The current status of the build. Valid values include:

FAILED: The build failed.

FAULT: The build faulted.

IN_PROGRESS: The build is still in progress.

STOPPED: The build stopped.

SUCCEEDED: The build succeeded.

TIMED_OUT: The build timed out.

build_endTime :: Lens' Build (Maybe UTCTime)

When the build process ended, expressed in Unix time format.

build_buildComplete :: Lens' Build (Maybe Bool)

Whether the build is complete. True if complete; otherwise, false.

build_logs :: Lens' Build (Maybe LogsLocation)

Information about the build's logs in CloudWatch Logs.

build_debugSession :: Lens' Build (Maybe DebugSession)

Contains information about the debug session for this build.

build_queuedTimeoutInMinutes :: Lens' Build (Maybe Int)

The number of minutes a build is allowed to be queued before it times out.

build_secondarySources :: Lens' Build (Maybe [ProjectSource])

An array of ProjectSource objects.

build_timeoutInMinutes :: Lens' Build (Maybe Int)

How long, in minutes, for CodeBuild to wait before timing out this build if it does not get marked as completed.

build_currentPhase :: Lens' Build (Maybe Text)

The current build phase.

build_initiator :: Lens' Build (Maybe Text)

The entity that started the build. Valid values include:

If CodePipeline started the build, the pipeline's name (for example, codepipeline/my-demo-pipeline).

If an IAM user started the build, the user's name (for example, MyUserName).

If the Jenkins plugin for CodeBuild started the build, the string CodeBuild-Jenkins-Plugin.

build_reportArns :: Lens' Build (Maybe [Text])

An array of the ARNs associated with this build's reports.

build_exportedEnvironmentVariables :: Lens' Build (Maybe [ExportedEnvironmentVariable])

A list of exported environment variables for this build.

Exported environment variables are used in conjunction with CodePipeline to export environment variables from the current build stage to subsequent stages in the pipeline. For more information, see Working with variables in the CodePipeline User Guide.

build_fileSystemLocations :: Lens' Build (Maybe [ProjectFileSystemLocation])

An array of ProjectFileSystemLocation objects for a CodeBuild build project. A ProjectFileSystemLocation object specifies the identifier, location, mountOptions, mountPoint, and type of a file system created using Amazon Elastic File System.