Network.AWS.CodeBuild.Types.BuildBatch

Description

Documentation

data BuildBatch

Contains information about a batch build.

See: newBuildBatch smart constructor.

Constructors

BuildBatch'

Fields

vpcConfig :: Maybe VpcConfig

sourceVersion :: Maybe Text

The identifier of the version of the source code to be built.

secondaryArtifacts :: Maybe [BuildArtifacts]

An array of BuildArtifacts objects the define the build artifacts for this batch build.

resolvedSourceVersion :: Maybe Text

The identifier of the resolved version of this batch build's source code.

For CodeCommit, GitHub, GitHub Enterprise, and BitBucket, the commit ID.

For CodePipeline, the source revision provided by CodePipeline.

For Amazon S3, this does not apply.

phases :: Maybe [BuildBatchPhase]

An array of BuildBatchPhase objects the specify the phases of the batch build.

cache :: Maybe ProjectCache

serviceRole :: Maybe Text

The name of a service role used for builds in the batch.

buildBatchNumber :: Maybe Integer

The number of the batch build. For each project, the buildBatchNumber of its first batch build is 1. The buildBatchNumber of each subsequent batch build is incremented by 1. If a batch build is deleted, the buildBatchNumber of other batch builds does not change.

secondarySourceVersions :: Maybe [ProjectSourceVersion]

An array of ProjectSourceVersion objects. Each ProjectSourceVersion must be one of:

For CodeCommit: the commit ID, branch, or Git tag to use.

For GitHub: the commit ID, pull request ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a pull request ID is specified, it must use the format pr/pull-request-ID (for example, pr/25). If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

For Bitbucket: the commit ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

For Amazon S3: the version ID of the object that represents the build input ZIP file to use.

encryptionKey :: Maybe Text

The Key Management Service customer master key (CMK) to be used for encrypting the batch build output artifacts.

You can use a cross-account KMS key to encrypt the build output artifacts if your service role has permission to that key.

You can specify either the Amazon Resource Name (ARN) of the CMK or, if available, the CMK's alias (using the format alias/<alias-name>).

id :: Maybe Text

The identifier of the batch build.

environment :: Maybe ProjectEnvironment

source :: Maybe ProjectSource

startTime :: Maybe POSIX

The date and time that the batch build started.

arn :: Maybe Text

The ARN of the batch build.

artifacts :: Maybe BuildArtifacts

A BuildArtifacts object the defines the build artifacts for this batch build.

projectName :: Maybe Text

The name of the batch build project.

endTime :: Maybe POSIX

The date and time that the batch build ended.

buildGroups :: Maybe [BuildGroup]

An array of BuildGroup objects that define the build groups for the batch build.

buildTimeoutInMinutes :: Maybe Int

Specifies the maximum amount of time, in minutes, that the build in a batch must be completed in.

queuedTimeoutInMinutes :: Maybe Int

Specifies the amount of time, in minutes, that the batch build is allowed to be queued before it times out.

secondarySources :: Maybe [ProjectSource]

An array of ProjectSource objects that define the sources for the batch build.

complete :: Maybe Bool

Indicates if the batch build is complete.

currentPhase :: Maybe Text

The current phase of the batch build.

buildBatchStatus :: Maybe StatusType

The status of the batch build.

logConfig :: Maybe LogsConfig

initiator :: Maybe Text

The entity that started the batch build. Valid values include:

If CodePipeline started the build, the pipeline's name (for example, codepipeline/my-demo-pipeline).

If an IAM user started the build, the user's name.

If the Jenkins plugin for CodeBuild started the build, the string CodeBuild-Jenkins-Plugin.

buildBatchConfig :: Maybe ProjectBuildBatchConfig

fileSystemLocations :: Maybe [ProjectFileSystemLocation]

An array of ProjectFileSystemLocation objects for the batch build project. A ProjectFileSystemLocation object specifies the identifier, location, mountOptions, mountPoint, and type of a file system created using Amazon Elastic File System.

debugSessionEnabled :: Maybe Bool

Specifies if session debugging is enabled for this batch build. For more information, see Viewing a running build in Session Manager. Batch session debugging is not supported for matrix batch builds.

newBuildBatch :: BuildBatch

Create a value of BuildBatch with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpcConfig:BuildBatch', buildBatch_vpcConfig - Undocumented member.

$sel:sourceVersion:BuildBatch', buildBatch_sourceVersion - The identifier of the version of the source code to be built.

$sel:secondaryArtifacts:BuildBatch', buildBatch_secondaryArtifacts - An array of BuildArtifacts objects the define the build artifacts for this batch build.

$sel:resolvedSourceVersion:BuildBatch', buildBatch_resolvedSourceVersion - The identifier of the resolved version of this batch build's source code.

For CodeCommit, GitHub, GitHub Enterprise, and BitBucket, the commit ID.

For CodePipeline, the source revision provided by CodePipeline.

For Amazon S3, this does not apply.

$sel:phases:BuildBatch', buildBatch_phases - An array of BuildBatchPhase objects the specify the phases of the batch build.

$sel:cache:BuildBatch', buildBatch_cache - Undocumented member.

$sel:serviceRole:BuildBatch', buildBatch_serviceRole - The name of a service role used for builds in the batch.

$sel:buildBatchNumber:BuildBatch', buildBatch_buildBatchNumber - The number of the batch build. For each project, the buildBatchNumber of its first batch build is 1. The buildBatchNumber of each subsequent batch build is incremented by 1. If a batch build is deleted, the buildBatchNumber of other batch builds does not change.

$sel:secondarySourceVersions:BuildBatch', buildBatch_secondarySourceVersions - An array of ProjectSourceVersion objects. Each ProjectSourceVersion must be one of:

For CodeCommit: the commit ID, branch, or Git tag to use.

For GitHub: the commit ID, pull request ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a pull request ID is specified, it must use the format pr/pull-request-ID (for example, pr/25). If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

For Bitbucket: the commit ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

For Amazon S3: the version ID of the object that represents the build input ZIP file to use.

$sel:encryptionKey:BuildBatch', buildBatch_encryptionKey - The Key Management Service customer master key (CMK) to be used for encrypting the batch build output artifacts.

You can use a cross-account KMS key to encrypt the build output artifacts if your service role has permission to that key.

You can specify either the Amazon Resource Name (ARN) of the CMK or, if available, the CMK's alias (using the format alias/<alias-name>).

$sel:id:BuildBatch', buildBatch_id - The identifier of the batch build.

$sel:environment:BuildBatch', buildBatch_environment - Undocumented member.

$sel:source:BuildBatch', buildBatch_source - Undocumented member.

$sel:startTime:BuildBatch', buildBatch_startTime - The date and time that the batch build started.

$sel:arn:BuildBatch', buildBatch_arn - The ARN of the batch build.

$sel:artifacts:BuildBatch', buildBatch_artifacts - A BuildArtifacts object the defines the build artifacts for this batch build.

$sel:projectName:BuildBatch', buildBatch_projectName - The name of the batch build project.

$sel:endTime:BuildBatch', buildBatch_endTime - The date and time that the batch build ended.

$sel:buildGroups:BuildBatch', buildBatch_buildGroups - An array of BuildGroup objects that define the build groups for the batch build.

$sel:buildTimeoutInMinutes:BuildBatch', buildBatch_buildTimeoutInMinutes - Specifies the maximum amount of time, in minutes, that the build in a batch must be completed in.

$sel:queuedTimeoutInMinutes:BuildBatch', buildBatch_queuedTimeoutInMinutes - Specifies the amount of time, in minutes, that the batch build is allowed to be queued before it times out.

$sel:secondarySources:BuildBatch', buildBatch_secondarySources - An array of ProjectSource objects that define the sources for the batch build.

$sel:complete:BuildBatch', buildBatch_complete - Indicates if the batch build is complete.

$sel:currentPhase:BuildBatch', buildBatch_currentPhase - The current phase of the batch build.

$sel:buildBatchStatus:BuildBatch', buildBatch_buildBatchStatus - The status of the batch build.

$sel:logConfig:BuildBatch', buildBatch_logConfig - Undocumented member.

$sel:initiator:BuildBatch', buildBatch_initiator - The entity that started the batch build. Valid values include:

If CodePipeline started the build, the pipeline's name (for example, codepipeline/my-demo-pipeline).

If an IAM user started the build, the user's name.

If the Jenkins plugin for CodeBuild started the build, the string CodeBuild-Jenkins-Plugin.

$sel:buildBatchConfig:BuildBatch', buildBatch_buildBatchConfig - Undocumented member.

$sel:fileSystemLocations:BuildBatch', buildBatch_fileSystemLocations - An array of ProjectFileSystemLocation objects for the batch build project. A ProjectFileSystemLocation object specifies the identifier, location, mountOptions, mountPoint, and type of a file system created using Amazon Elastic File System.

$sel:debugSessionEnabled:BuildBatch', buildBatch_debugSessionEnabled - Specifies if session debugging is enabled for this batch build. For more information, see Viewing a running build in Session Manager. Batch session debugging is not supported for matrix batch builds.

buildBatch_vpcConfig :: Lens' BuildBatch (Maybe VpcConfig)

Undocumented member.

buildBatch_sourceVersion :: Lens' BuildBatch (Maybe Text)

The identifier of the version of the source code to be built.

buildBatch_secondaryArtifacts :: Lens' BuildBatch (Maybe [BuildArtifacts])

An array of BuildArtifacts objects the define the build artifacts for this batch build.

buildBatch_resolvedSourceVersion :: Lens' BuildBatch (Maybe Text)

The identifier of the resolved version of this batch build's source code.

For CodeCommit, GitHub, GitHub Enterprise, and BitBucket, the commit ID.

For CodePipeline, the source revision provided by CodePipeline.

For Amazon S3, this does not apply.

buildBatch_phases :: Lens' BuildBatch (Maybe [BuildBatchPhase])

An array of BuildBatchPhase objects the specify the phases of the batch build.

buildBatch_cache :: Lens' BuildBatch (Maybe ProjectCache)

Undocumented member.

buildBatch_serviceRole :: Lens' BuildBatch (Maybe Text)

The name of a service role used for builds in the batch.

buildBatch_buildBatchNumber :: Lens' BuildBatch (Maybe Integer)

The number of the batch build. For each project, the buildBatchNumber of its first batch build is 1. The buildBatchNumber of each subsequent batch build is incremented by 1. If a batch build is deleted, the buildBatchNumber of other batch builds does not change.

buildBatch_secondarySourceVersions :: Lens' BuildBatch (Maybe [ProjectSourceVersion])

An array of ProjectSourceVersion objects. Each ProjectSourceVersion must be one of:

For CodeCommit: the commit ID, branch, or Git tag to use.

For GitHub: the commit ID, pull request ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a pull request ID is specified, it must use the format pr/pull-request-ID (for example, pr/25). If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

For Bitbucket: the commit ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

For Amazon S3: the version ID of the object that represents the build input ZIP file to use.

buildBatch_encryptionKey :: Lens' BuildBatch (Maybe Text)

The Key Management Service customer master key (CMK) to be used for encrypting the batch build output artifacts.

You can use a cross-account KMS key to encrypt the build output artifacts if your service role has permission to that key.

You can specify either the Amazon Resource Name (ARN) of the CMK or, if available, the CMK's alias (using the format alias/<alias-name>).

buildBatch_id :: Lens' BuildBatch (Maybe Text)

The identifier of the batch build.

buildBatch_environment :: Lens' BuildBatch (Maybe ProjectEnvironment)

Undocumented member.

buildBatch_source :: Lens' BuildBatch (Maybe ProjectSource)

Undocumented member.

buildBatch_startTime :: Lens' BuildBatch (Maybe UTCTime)

The date and time that the batch build started.

buildBatch_arn :: Lens' BuildBatch (Maybe Text)

The ARN of the batch build.

buildBatch_artifacts :: Lens' BuildBatch (Maybe BuildArtifacts)

A BuildArtifacts object the defines the build artifacts for this batch build.

buildBatch_projectName :: Lens' BuildBatch (Maybe Text)

The name of the batch build project.

buildBatch_endTime :: Lens' BuildBatch (Maybe UTCTime)

The date and time that the batch build ended.

buildBatch_buildGroups :: Lens' BuildBatch (Maybe [BuildGroup])

An array of BuildGroup objects that define the build groups for the batch build.

buildBatch_buildTimeoutInMinutes :: Lens' BuildBatch (Maybe Int)

Specifies the maximum amount of time, in minutes, that the build in a batch must be completed in.

buildBatch_queuedTimeoutInMinutes :: Lens' BuildBatch (Maybe Int)

Specifies the amount of time, in minutes, that the batch build is allowed to be queued before it times out.

buildBatch_secondarySources :: Lens' BuildBatch (Maybe [ProjectSource])

An array of ProjectSource objects that define the sources for the batch build.

buildBatch_complete :: Lens' BuildBatch (Maybe Bool)

Indicates if the batch build is complete.

buildBatch_currentPhase :: Lens' BuildBatch (Maybe Text)

The current phase of the batch build.

buildBatch_buildBatchStatus :: Lens' BuildBatch (Maybe StatusType)

The status of the batch build.

buildBatch_logConfig :: Lens' BuildBatch (Maybe LogsConfig)

Undocumented member.

buildBatch_initiator :: Lens' BuildBatch (Maybe Text)

The entity that started the batch build. Valid values include:

If CodePipeline started the build, the pipeline's name (for example, codepipeline/my-demo-pipeline).

If an IAM user started the build, the user's name.

If the Jenkins plugin for CodeBuild started the build, the string CodeBuild-Jenkins-Plugin.

buildBatch_buildBatchConfig :: Lens' BuildBatch (Maybe ProjectBuildBatchConfig)

Undocumented member.

buildBatch_fileSystemLocations :: Lens' BuildBatch (Maybe [ProjectFileSystemLocation])

An array of ProjectFileSystemLocation objects for the batch build project. A ProjectFileSystemLocation object specifies the identifier, location, mountOptions, mountPoint, and type of a file system created using Amazon Elastic File System.

buildBatch_debugSessionEnabled :: Lens' BuildBatch (Maybe Bool)

Specifies if session debugging is enabled for this batch build. For more information, see Viewing a running build in Session Manager. Batch session debugging is not supported for matrix batch builds.