Network.AWS.CodeBuild.Types.BuildBatchPhase

Description

Documentation

data BuildBatchPhase

Contains information about a stage for a batch build.

See: newBuildBatchPhase smart constructor.

Constructors

BuildBatchPhase'

Fields

phaseType :: Maybe BuildBatchPhaseType

The name of the batch build phase. Valid values include:

COMBINE_ARTIFACTS

Build output artifacts are being combined and uploaded to the output location.

DOWNLOAD_BATCHSPEC

The batch build specification is being downloaded.

FAILED

One or more of the builds failed.

IN_PROGRESS

The batch build is in progress.

STOPPED

The batch build was stopped.

SUBMITTED

The btach build has been submitted.

SUCCEEDED

The batch build succeeded.

contexts :: Maybe [PhaseContext]

Additional information about the batch build phase. Especially to help troubleshoot a failed batch build.

startTime :: Maybe POSIX

When the batch build phase started, expressed in Unix time format.

endTime :: Maybe POSIX

When the batch build phase ended, expressed in Unix time format.

durationInSeconds :: Maybe Integer

How long, in seconds, between the starting and ending times of the batch build's phase.

phaseStatus :: Maybe StatusType

The current status of the batch build phase. Valid values include:

FAILED

The build phase failed.

FAULT

The build phase faulted.

IN_PROGRESS

The build phase is still in progress.

STOPPED

The build phase stopped.

SUCCEEDED

The build phase succeeded.

TIMED_OUT

The build phase timed out.

newBuildBatchPhase :: BuildBatchPhase

Create a value of BuildBatchPhase with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:phaseType:BuildBatchPhase', buildBatchPhase_phaseType - The name of the batch build phase. Valid values include:

COMBINE_ARTIFACTS

Build output artifacts are being combined and uploaded to the output location.

DOWNLOAD_BATCHSPEC

The batch build specification is being downloaded.

FAILED

One or more of the builds failed.

IN_PROGRESS

The batch build is in progress.

STOPPED

The batch build was stopped.

SUBMITTED

The btach build has been submitted.

SUCCEEDED

The batch build succeeded.

$sel:contexts:BuildBatchPhase', buildBatchPhase_contexts - Additional information about the batch build phase. Especially to help troubleshoot a failed batch build.

$sel:startTime:BuildBatchPhase', buildBatchPhase_startTime - When the batch build phase started, expressed in Unix time format.

$sel:endTime:BuildBatchPhase', buildBatchPhase_endTime - When the batch build phase ended, expressed in Unix time format.

$sel:durationInSeconds:BuildBatchPhase', buildBatchPhase_durationInSeconds - How long, in seconds, between the starting and ending times of the batch build's phase.

$sel:phaseStatus:BuildBatchPhase', buildBatchPhase_phaseStatus - The current status of the batch build phase. Valid values include:

FAILED

The build phase failed.

FAULT

The build phase faulted.

IN_PROGRESS

The build phase is still in progress.

STOPPED

The build phase stopped.

SUCCEEDED

The build phase succeeded.

TIMED_OUT

The build phase timed out.

buildBatchPhase_phaseType :: Lens' BuildBatchPhase (Maybe BuildBatchPhaseType)

The name of the batch build phase. Valid values include:

COMBINE_ARTIFACTS

Build output artifacts are being combined and uploaded to the output location.

DOWNLOAD_BATCHSPEC

The batch build specification is being downloaded.

FAILED

One or more of the builds failed.

IN_PROGRESS

The batch build is in progress.

STOPPED

The batch build was stopped.

SUBMITTED

The btach build has been submitted.

SUCCEEDED

The batch build succeeded.

buildBatchPhase_contexts :: Lens' BuildBatchPhase (Maybe [PhaseContext])

Additional information about the batch build phase. Especially to help troubleshoot a failed batch build.

buildBatchPhase_startTime :: Lens' BuildBatchPhase (Maybe UTCTime)

When the batch build phase started, expressed in Unix time format.

buildBatchPhase_endTime :: Lens' BuildBatchPhase (Maybe UTCTime)

When the batch build phase ended, expressed in Unix time format.

buildBatchPhase_durationInSeconds :: Lens' BuildBatchPhase (Maybe Integer)

How long, in seconds, between the starting and ending times of the batch build's phase.

buildBatchPhase_phaseStatus :: Lens' BuildBatchPhase (Maybe StatusType)

The current status of the batch build phase. Valid values include:

FAILED

The build phase failed.

FAULT

The build phase faulted.

IN_PROGRESS

The build phase is still in progress.

STOPPED

The build phase stopped.

SUCCEEDED

The build phase succeeded.

TIMED_OUT

The build phase timed out.