Network.AWS.CodeBuild.Types.BuildGroup

Description

Documentation

data BuildGroup

Contains information about a batch build build group. Build groups are used to combine builds that can run in parallel, while still being able to set dependencies on other build groups.

See: newBuildGroup smart constructor.

Constructors

BuildGroup'

Fields

identifier :: Maybe Text

Contains the identifier of the build group.

dependsOn :: Maybe [Text]

An array of strings that contain the identifiers of the build groups that this build group depends on.

ignoreFailure :: Maybe Bool

Specifies if failures in this build group can be ignored.

currentBuildSummary :: Maybe BuildSummary

A BuildSummary object that contains a summary of the current build group.

priorBuildSummaryList :: Maybe [BuildSummary]

An array of BuildSummary objects that contain summaries of previous build groups.

newBuildGroup :: BuildGroup

Create a value of BuildGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identifier:BuildGroup', buildGroup_identifier - Contains the identifier of the build group.

$sel:dependsOn:BuildGroup', buildGroup_dependsOn - An array of strings that contain the identifiers of the build groups that this build group depends on.

$sel:ignoreFailure:BuildGroup', buildGroup_ignoreFailure - Specifies if failures in this build group can be ignored.

$sel:currentBuildSummary:BuildGroup', buildGroup_currentBuildSummary - A BuildSummary object that contains a summary of the current build group.

$sel:priorBuildSummaryList:BuildGroup', buildGroup_priorBuildSummaryList - An array of BuildSummary objects that contain summaries of previous build groups.

buildGroup_identifier :: Lens' BuildGroup (Maybe Text)

Contains the identifier of the build group.

buildGroup_dependsOn :: Lens' BuildGroup (Maybe [Text])

An array of strings that contain the identifiers of the build groups that this build group depends on.

buildGroup_ignoreFailure :: Lens' BuildGroup (Maybe Bool)

Specifies if failures in this build group can be ignored.

buildGroup_currentBuildSummary :: Lens' BuildGroup (Maybe BuildSummary)

A BuildSummary object that contains a summary of the current build group.

buildGroup_priorBuildSummaryList :: Lens' BuildGroup (Maybe [BuildSummary])

An array of BuildSummary objects that contain summaries of previous build groups.