Network.AWS.CodeBuild.Types.BuildPhase

Description

Documentation

data BuildPhase

Information about a stage for a build.

See: newBuildPhase smart constructor.

Constructors

BuildPhase'

Fields

phaseType :: Maybe BuildPhaseType

The name of the build phase. Valid values include:

BUILD

Core build activities typically occur in this build phase.

COMPLETED

The build has been completed.

DOWNLOAD_SOURCE

Source code is being downloaded in this build phase.

FINALIZING

The build process is completing in this build phase.

INSTALL

Installation activities typically occur in this build phase.

POST_BUILD

Post-build activities typically occur in this build phase.

PRE_BUILD

Pre-build activities typically occur in this build phase.

PROVISIONING

The build environment is being set up.

QUEUED

The build has been submitted and is queued behind other submitted builds.

SUBMITTED

The build has been submitted.

UPLOAD_ARTIFACTS

Build output artifacts are being uploaded to the output location.

contexts :: Maybe [PhaseContext]

Additional information about a build phase, especially to help troubleshoot a failed build.

startTime :: Maybe POSIX

When the build phase started, expressed in Unix time format.

endTime :: Maybe POSIX

When the build phase ended, expressed in Unix time format.

durationInSeconds :: Maybe Integer

How long, in seconds, between the starting and ending times of the build's phase.

phaseStatus :: Maybe StatusType

The current status of the build phase. Valid values include:

FAILED

The build phase failed.

FAULT

The build phase faulted.

IN_PROGRESS

The build phase is still in progress.

STOPPED

The build phase stopped.

SUCCEEDED

The build phase succeeded.

TIMED_OUT

The build phase timed out.

newBuildPhase :: BuildPhase

Create a value of BuildPhase with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:phaseType:BuildPhase', buildPhase_phaseType - The name of the build phase. Valid values include:

BUILD

Core build activities typically occur in this build phase.

COMPLETED

The build has been completed.

DOWNLOAD_SOURCE

Source code is being downloaded in this build phase.

FINALIZING

The build process is completing in this build phase.

INSTALL

Installation activities typically occur in this build phase.

POST_BUILD

Post-build activities typically occur in this build phase.

PRE_BUILD

Pre-build activities typically occur in this build phase.

PROVISIONING

The build environment is being set up.

QUEUED

The build has been submitted and is queued behind other submitted builds.

SUBMITTED

The build has been submitted.

UPLOAD_ARTIFACTS

Build output artifacts are being uploaded to the output location.

$sel:contexts:BuildPhase', buildPhase_contexts - Additional information about a build phase, especially to help troubleshoot a failed build.

$sel:startTime:BuildPhase', buildPhase_startTime - When the build phase started, expressed in Unix time format.

$sel:endTime:BuildPhase', buildPhase_endTime - When the build phase ended, expressed in Unix time format.

$sel:durationInSeconds:BuildPhase', buildPhase_durationInSeconds - How long, in seconds, between the starting and ending times of the build's phase.

$sel:phaseStatus:BuildPhase', buildPhase_phaseStatus - The current status of the build phase. Valid values include:

FAILED

The build phase failed.

FAULT

The build phase faulted.

IN_PROGRESS

The build phase is still in progress.

STOPPED

The build phase stopped.

SUCCEEDED

The build phase succeeded.

TIMED_OUT

The build phase timed out.

buildPhase_phaseType :: Lens' BuildPhase (Maybe BuildPhaseType)

The name of the build phase. Valid values include:

BUILD

Core build activities typically occur in this build phase.

COMPLETED

The build has been completed.

DOWNLOAD_SOURCE

Source code is being downloaded in this build phase.

FINALIZING

The build process is completing in this build phase.

INSTALL

Installation activities typically occur in this build phase.

POST_BUILD

Post-build activities typically occur in this build phase.

PRE_BUILD

Pre-build activities typically occur in this build phase.

PROVISIONING

The build environment is being set up.

QUEUED

The build has been submitted and is queued behind other submitted builds.

SUBMITTED

The build has been submitted.

UPLOAD_ARTIFACTS

Build output artifacts are being uploaded to the output location.

buildPhase_contexts :: Lens' BuildPhase (Maybe [PhaseContext])

Additional information about a build phase, especially to help troubleshoot a failed build.

buildPhase_startTime :: Lens' BuildPhase (Maybe UTCTime)

When the build phase started, expressed in Unix time format.

buildPhase_endTime :: Lens' BuildPhase (Maybe UTCTime)

When the build phase ended, expressed in Unix time format.

buildPhase_durationInSeconds :: Lens' BuildPhase (Maybe Integer)

How long, in seconds, between the starting and ending times of the build's phase.

buildPhase_phaseStatus :: Lens' BuildPhase (Maybe StatusType)

The current status of the build phase. Valid values include:

FAILED

The build phase failed.

FAULT

The build phase faulted.

IN_PROGRESS

The build phase is still in progress.

STOPPED

The build phase stopped.

SUCCEEDED

The build phase succeeded.

TIMED_OUT

The build phase timed out.