Network.AWS.CodeBuild.Types.BuildSummary

Description

Documentation

data BuildSummary

Contains summary information about a batch build group.

See: newBuildSummary smart constructor.

Constructors

BuildSummary'

Fields

secondaryArtifacts :: Maybe [ResolvedArtifact]

An array of ResolvedArtifact objects that represents the secondary build artifacts for the build group.

primaryArtifact :: Maybe ResolvedArtifact

A ResolvedArtifact object that represents the primary build artifacts for the build group.

arn :: Maybe Text

The batch build ARN.

buildStatus :: Maybe StatusType

The status of the build group.

FAILED

The build group failed.

FAULT

The build group faulted.

IN_PROGRESS

The build group is still in progress.

STOPPED

The build group stopped.

SUCCEEDED

The build group succeeded.

TIMED_OUT

The build group timed out.

requestedOn :: Maybe POSIX

When the build was started, expressed in Unix time format.

newBuildSummary :: BuildSummary

Create a value of BuildSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:secondaryArtifacts:BuildSummary', buildSummary_secondaryArtifacts - An array of ResolvedArtifact objects that represents the secondary build artifacts for the build group.

$sel:primaryArtifact:BuildSummary', buildSummary_primaryArtifact - A ResolvedArtifact object that represents the primary build artifacts for the build group.

$sel:arn:BuildSummary', buildSummary_arn - The batch build ARN.

$sel:buildStatus:BuildSummary', buildSummary_buildStatus - The status of the build group.

FAILED

The build group failed.

FAULT

The build group faulted.

IN_PROGRESS

The build group is still in progress.

STOPPED

The build group stopped.

SUCCEEDED

The build group succeeded.

TIMED_OUT

The build group timed out.

$sel:requestedOn:BuildSummary', buildSummary_requestedOn - When the build was started, expressed in Unix time format.

buildSummary_secondaryArtifacts :: Lens' BuildSummary (Maybe [ResolvedArtifact])

An array of ResolvedArtifact objects that represents the secondary build artifacts for the build group.

buildSummary_primaryArtifact :: Lens' BuildSummary (Maybe ResolvedArtifact)

A ResolvedArtifact object that represents the primary build artifacts for the build group.

buildSummary_arn :: Lens' BuildSummary (Maybe Text)

The batch build ARN.

buildSummary_buildStatus :: Lens' BuildSummary (Maybe StatusType)

The status of the build group.

FAILED

The build group failed.

FAULT

The build group faulted.

IN_PROGRESS

The build group is still in progress.

STOPPED

The build group stopped.

SUCCEEDED

The build group succeeded.

TIMED_OUT

The build group timed out.

buildSummary_requestedOn :: Lens' BuildSummary (Maybe UTCTime)

When the build was started, expressed in Unix time format.