Network.AWS.CodeBuild.Types.CodeCoverage

Description

Documentation

data CodeCoverage

Contains code coverage report information.

Line coverage measures how many statements your tests cover. A statement is a single instruction, not including comments, conditionals, etc.

Branch coverage determines if your tests cover every possible branch of a control structure, such as an if or case statement.

See: newCodeCoverage smart constructor.

Constructors

CodeCoverage'

Fields

branchesMissed :: Maybe Natural

The number of conditional branches that are not covered by your tests.

linesCovered :: Maybe Natural

The number of lines that are covered by your tests.

branchesCovered :: Maybe Natural

The number of conditional branches that are covered by your tests.

filePath :: Maybe Text

The path of the test report file.

reportARN :: Maybe Text

The ARN of the report.

id :: Maybe Text

The identifier of the code coverage report.

expired :: Maybe POSIX

The date and time that the tests were run.

lineCoveragePercentage :: Maybe Double

The percentage of lines that are covered by your tests.

linesMissed :: Maybe Natural

The number of lines that are not covered by your tests.

branchCoveragePercentage :: Maybe Double

The percentage of branches that are covered by your tests.

newCodeCoverage :: CodeCoverage

Create a value of CodeCoverage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:branchesMissed:CodeCoverage', codeCoverage_branchesMissed - The number of conditional branches that are not covered by your tests.

$sel:linesCovered:CodeCoverage', codeCoverage_linesCovered - The number of lines that are covered by your tests.

$sel:branchesCovered:CodeCoverage', codeCoverage_branchesCovered - The number of conditional branches that are covered by your tests.

$sel:filePath:CodeCoverage', codeCoverage_filePath - The path of the test report file.

$sel:reportARN:CodeCoverage', codeCoverage_reportARN - The ARN of the report.

$sel:id:CodeCoverage', codeCoverage_id - The identifier of the code coverage report.

$sel:expired:CodeCoverage', codeCoverage_expired - The date and time that the tests were run.

$sel:lineCoveragePercentage:CodeCoverage', codeCoverage_lineCoveragePercentage - The percentage of lines that are covered by your tests.

$sel:linesMissed:CodeCoverage', codeCoverage_linesMissed - The number of lines that are not covered by your tests.

$sel:branchCoveragePercentage:CodeCoverage', codeCoverage_branchCoveragePercentage - The percentage of branches that are covered by your tests.

codeCoverage_branchesMissed :: Lens' CodeCoverage (Maybe Natural)

The number of conditional branches that are not covered by your tests.

codeCoverage_linesCovered :: Lens' CodeCoverage (Maybe Natural)

The number of lines that are covered by your tests.

codeCoverage_branchesCovered :: Lens' CodeCoverage (Maybe Natural)

The number of conditional branches that are covered by your tests.

codeCoverage_filePath :: Lens' CodeCoverage (Maybe Text)

The path of the test report file.

codeCoverage_reportARN :: Lens' CodeCoverage (Maybe Text)

The ARN of the report.

codeCoverage_id :: Lens' CodeCoverage (Maybe Text)

The identifier of the code coverage report.

codeCoverage_expired :: Lens' CodeCoverage (Maybe UTCTime)

The date and time that the tests were run.

codeCoverage_lineCoveragePercentage :: Lens' CodeCoverage (Maybe Double)

The percentage of lines that are covered by your tests.

codeCoverage_linesMissed :: Lens' CodeCoverage (Maybe Natural)

The number of lines that are not covered by your tests.

codeCoverage_branchCoveragePercentage :: Lens' CodeCoverage (Maybe Double)

The percentage of branches that are covered by your tests.