Network.AWS.CodeBuild.Types.CodeCoverageReportSummary

Description

Documentation

data CodeCoverageReportSummary

Contains a summary of a code coverage report.

Line coverage measures how many statements your tests cover. A statement is a single instruction, not including comments, conditionals, etc.

Branch coverage determines if your tests cover every possible branch of a control structure, such as an if or case statement.

See: newCodeCoverageReportSummary smart constructor.

Constructors

CodeCoverageReportSummary'

Fields

branchesMissed :: Maybe Natural

The number of conditional branches that are not covered by your tests.

linesCovered :: Maybe Natural

The number of lines that are covered by your tests.

branchesCovered :: Maybe Natural

The number of conditional branches that are covered by your tests.

lineCoveragePercentage :: Maybe Double

The percentage of lines that are covered by your tests.

linesMissed :: Maybe Natural

The number of lines that are not covered by your tests.

branchCoveragePercentage :: Maybe Double

The percentage of branches that are covered by your tests.

newCodeCoverageReportSummary :: CodeCoverageReportSummary

Create a value of CodeCoverageReportSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:branchesMissed:CodeCoverageReportSummary', codeCoverageReportSummary_branchesMissed - The number of conditional branches that are not covered by your tests.

$sel:linesCovered:CodeCoverageReportSummary', codeCoverageReportSummary_linesCovered - The number of lines that are covered by your tests.

$sel:branchesCovered:CodeCoverageReportSummary', codeCoverageReportSummary_branchesCovered - The number of conditional branches that are covered by your tests.

$sel:lineCoveragePercentage:CodeCoverageReportSummary', codeCoverageReportSummary_lineCoveragePercentage - The percentage of lines that are covered by your tests.

$sel:linesMissed:CodeCoverageReportSummary', codeCoverageReportSummary_linesMissed - The number of lines that are not covered by your tests.

$sel:branchCoveragePercentage:CodeCoverageReportSummary', codeCoverageReportSummary_branchCoveragePercentage - The percentage of branches that are covered by your tests.

codeCoverageReportSummary_branchesMissed :: Lens' CodeCoverageReportSummary (Maybe Natural)

The number of conditional branches that are not covered by your tests.

codeCoverageReportSummary_linesCovered :: Lens' CodeCoverageReportSummary (Maybe Natural)

The number of lines that are covered by your tests.

codeCoverageReportSummary_branchesCovered :: Lens' CodeCoverageReportSummary (Maybe Natural)

The number of conditional branches that are covered by your tests.

codeCoverageReportSummary_lineCoveragePercentage :: Lens' CodeCoverageReportSummary (Maybe Double)

The percentage of lines that are covered by your tests.

codeCoverageReportSummary_linesMissed :: Lens' CodeCoverageReportSummary (Maybe Natural)

The number of lines that are not covered by your tests.

codeCoverageReportSummary_branchCoveragePercentage :: Lens' CodeCoverageReportSummary (Maybe Double)

The percentage of branches that are covered by your tests.