Network.AWS.CodeBuild.Types.DebugSession

Description

Documentation

data DebugSession

Contains information about the debug session for a build. For more information, see Viewing a running build in Session Manager.

See: newDebugSession smart constructor.

Constructors

DebugSession'

Fields

sessionTarget :: Maybe Text

Contains the identifier of the Session Manager session used for the build. To work with the paused build, you open this session to examine, control, and resume the build.

sessionEnabled :: Maybe Bool

Specifies if session debugging is enabled for this build.

newDebugSession :: DebugSession

Create a value of DebugSession with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sessionTarget:DebugSession', debugSession_sessionTarget - Contains the identifier of the Session Manager session used for the build. To work with the paused build, you open this session to examine, control, and resume the build.

$sel:sessionEnabled:DebugSession', debugSession_sessionEnabled - Specifies if session debugging is enabled for this build.

debugSession_sessionTarget :: Lens' DebugSession (Maybe Text)

Contains the identifier of the Session Manager session used for the build. To work with the paused build, you open this session to examine, control, and resume the build.

debugSession_sessionEnabled :: Lens' DebugSession (Maybe Bool)

Specifies if session debugging is enabled for this build.