Network.AWS.CodeBuild.Types.LogsLocation

Description

Documentation

data LogsLocation

Information about build logs in CloudWatch Logs.

See: newLogsLocation smart constructor.

Constructors

LogsLocation'

Fields

deepLink :: Maybe Text

The URL to an individual build log in CloudWatch Logs.

s3Logs :: Maybe S3LogsConfig

Information about S3 logs for a build project.

cloudWatchLogs :: Maybe CloudWatchLogsConfig

Information about CloudWatch Logs for a build project.

s3DeepLink :: Maybe Text

The URL to a build log in an S3 bucket.

s3LogsArn :: Maybe Text

The ARN of S3 logs for a build project. Its format is arn:${Partition}:s3:::${BucketName}/${ObjectName}. For more information, see Resources Defined by Amazon S3.

cloudWatchLogsArn :: Maybe Text

The ARN of CloudWatch Logs for a build project. Its format is arn:${Partition}:logs:${Region}:${Account}:log-group:${LogGroupName}:log-stream:${LogStreamName}. For more information, see Resources Defined by CloudWatch Logs.

groupName :: Maybe Text

The name of the CloudWatch Logs group for the build logs.

streamName :: Maybe Text

The name of the CloudWatch Logs stream for the build logs.

newLogsLocation :: LogsLocation

Create a value of LogsLocation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deepLink:LogsLocation', logsLocation_deepLink - The URL to an individual build log in CloudWatch Logs.

$sel:s3Logs:LogsLocation', logsLocation_s3Logs - Information about S3 logs for a build project.

$sel:cloudWatchLogs:LogsLocation', logsLocation_cloudWatchLogs - Information about CloudWatch Logs for a build project.

$sel:s3DeepLink:LogsLocation', logsLocation_s3DeepLink - The URL to a build log in an S3 bucket.

$sel:s3LogsArn:LogsLocation', logsLocation_s3LogsArn - The ARN of S3 logs for a build project. Its format is arn:${Partition}:s3:::${BucketName}/${ObjectName}. For more information, see Resources Defined by Amazon S3.

$sel:cloudWatchLogsArn:LogsLocation', logsLocation_cloudWatchLogsArn - The ARN of CloudWatch Logs for a build project. Its format is arn:${Partition}:logs:${Region}:${Account}:log-group:${LogGroupName}:log-stream:${LogStreamName}. For more information, see Resources Defined by CloudWatch Logs.

$sel:groupName:LogsLocation', logsLocation_groupName - The name of the CloudWatch Logs group for the build logs.

$sel:streamName:LogsLocation', logsLocation_streamName - The name of the CloudWatch Logs stream for the build logs.

logsLocation_deepLink :: Lens' LogsLocation (Maybe Text)

The URL to an individual build log in CloudWatch Logs.

logsLocation_s3Logs :: Lens' LogsLocation (Maybe S3LogsConfig)

Information about S3 logs for a build project.

logsLocation_cloudWatchLogs :: Lens' LogsLocation (Maybe CloudWatchLogsConfig)

Information about CloudWatch Logs for a build project.

logsLocation_s3DeepLink :: Lens' LogsLocation (Maybe Text)

The URL to a build log in an S3 bucket.

logsLocation_s3LogsArn :: Lens' LogsLocation (Maybe Text)

The ARN of S3 logs for a build project. Its format is arn:${Partition}:s3:::${BucketName}/${ObjectName}. For more information, see Resources Defined by Amazon S3.

logsLocation_cloudWatchLogsArn :: Lens' LogsLocation (Maybe Text)

The ARN of CloudWatch Logs for a build project. Its format is arn:${Partition}:logs:${Region}:${Account}:log-group:${LogGroupName}:log-stream:${LogStreamName}. For more information, see Resources Defined by CloudWatch Logs.

logsLocation_groupName :: Lens' LogsLocation (Maybe Text)

The name of the CloudWatch Logs group for the build logs.

logsLocation_streamName :: Lens' LogsLocation (Maybe Text)

The name of the CloudWatch Logs stream for the build logs.