Network.AWS.CodeBuild.Types.Project

Description

Documentation

data Project

Information about a build project.

See: newProject smart constructor.

Constructors

Project'

Fields

secondaryArtifacts :: Maybe [ProjectArtifacts]

An array of ProjectArtifacts objects.

resourceAccessRole :: Maybe Text

The ARN of the IAM role that enables CodeBuild to access the CloudWatch Logs and Amazon S3 artifacts for the project's builds.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the build project.

artifacts :: Maybe ProjectArtifacts

Information about the build output artifacts for the build project.

environment :: Maybe ProjectEnvironment

Information about the build environment for this build project.

created :: Maybe POSIX

When the build project was created, expressed in Unix time format.

concurrentBuildLimit :: Maybe Int

The maximum number of concurrent builds that are allowed for this project.

New builds are only started if the current number of builds is less than or equal to this limit. If the current build count meets this limit, new builds are throttled and are not run.

secondarySourceVersions :: Maybe [ProjectSourceVersion]

An array of ProjectSourceVersion objects. If secondarySourceVersions is specified at the build level, then they take over these secondarySourceVersions (at the project level).

queuedTimeoutInMinutes :: Maybe Natural

The number of minutes a build is allowed to be queued before it times out.

cache :: Maybe ProjectCache

Information about the cache for the build project.

secondarySources :: Maybe [ProjectSource]

An array of ProjectSource objects.

sourceVersion :: Maybe Text

A version of the build input to be built for this project. If not specified, the latest version is used. If specified, it must be one of:

For CodeCommit: the commit ID, branch, or Git tag to use.

For GitHub: the commit ID, pull request ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a pull request ID is specified, it must use the format pr/pull-request-ID (for example pr/25). If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

For Bitbucket: the commit ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

For Amazon S3: the version ID of the object that represents the build input ZIP file to use.

If sourceVersion is specified at the build level, then that version takes precedence over this sourceVersion (at the project level).

For more information, see Source Version Sample with CodeBuild in the CodeBuild User Guide.

name :: Maybe Text

The name of the build project.

vpcConfig :: Maybe VpcConfig

Information about the VPC configuration that CodeBuild accesses.

publicProjectAlias :: Maybe Text

Contains the project identifier used with the public build APIs.

source :: Maybe ProjectSource

Information about the build input source code for this build project.

badge :: Maybe ProjectBadge

Information about the build badge for the build project.

logsConfig :: Maybe LogsConfig

Information about logs for the build project. A project can create logs in CloudWatch Logs, an S3 bucket, or both.

fileSystemLocations :: Maybe [ProjectFileSystemLocation]

An array of ProjectFileSystemLocation objects for a CodeBuild build project. A ProjectFileSystemLocation object specifies the identifier, location, mountOptions, mountPoint, and type of a file system created using Amazon Elastic File System.

buildBatchConfig :: Maybe ProjectBuildBatchConfig

A ProjectBuildBatchConfig object that defines the batch build options for the project.

encryptionKey :: Maybe Text

The Key Management Service customer master key (CMK) to be used for encrypting the build output artifacts.

You can use a cross-account KMS key to encrypt the build output artifacts if your service role has permission to that key.

You can specify either the Amazon Resource Name (ARN) of the CMK or, if available, the CMK's alias (using the format alias/<alias-name>). If you don't specify a value, CodeBuild uses the managed CMK for Amazon Simple Storage Service (Amazon S3).

lastModified :: Maybe POSIX

When the build project's settings were last modified, expressed in Unix time format.

projectVisibility :: Maybe ProjectVisibilityType

webhook :: Maybe Webhook

Information about a webhook that connects repository events to a build project in CodeBuild.

description :: Maybe Text

A description that makes the build project easy to identify.

serviceRole :: Maybe Text

The ARN of the IAM role that enables CodeBuild to interact with dependent Amazon Web Services services on behalf of the Amazon Web Services account.

tags :: Maybe [Tag]

A list of tag key and value pairs associated with this build project.

These tags are available for use by Amazon Web Services services that support CodeBuild build project tags.

timeoutInMinutes :: Maybe Natural

How long, in minutes, from 5 to 480 (8 hours), for CodeBuild to wait before timing out any related build that did not get marked as completed. The default is 60 minutes.

newProject :: Project

Create a value of Project with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:secondaryArtifacts:Project', project_secondaryArtifacts - An array of ProjectArtifacts objects.

$sel:resourceAccessRole:Project', project_resourceAccessRole - The ARN of the IAM role that enables CodeBuild to access the CloudWatch Logs and Amazon S3 artifacts for the project's builds.

$sel:arn:Project', project_arn - The Amazon Resource Name (ARN) of the build project.

$sel:artifacts:Project', project_artifacts - Information about the build output artifacts for the build project.

$sel:environment:Project', project_environment - Information about the build environment for this build project.

$sel:created:Project', project_created - When the build project was created, expressed in Unix time format.

$sel:concurrentBuildLimit:Project', project_concurrentBuildLimit - The maximum number of concurrent builds that are allowed for this project.

New builds are only started if the current number of builds is less than or equal to this limit. If the current build count meets this limit, new builds are throttled and are not run.

$sel:secondarySourceVersions:Project', project_secondarySourceVersions - An array of ProjectSourceVersion objects. If secondarySourceVersions is specified at the build level, then they take over these secondarySourceVersions (at the project level).

$sel:queuedTimeoutInMinutes:Project', project_queuedTimeoutInMinutes - The number of minutes a build is allowed to be queued before it times out.

$sel:cache:Project', project_cache - Information about the cache for the build project.

$sel:secondarySources:Project', project_secondarySources - An array of ProjectSource objects.

$sel:sourceVersion:Project', project_sourceVersion - A version of the build input to be built for this project. If not specified, the latest version is used. If specified, it must be one of:

For CodeCommit: the commit ID, branch, or Git tag to use.

For GitHub: the commit ID, pull request ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a pull request ID is specified, it must use the format pr/pull-request-ID (for example pr/25). If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

For Bitbucket: the commit ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

For Amazon S3: the version ID of the object that represents the build input ZIP file to use.

If sourceVersion is specified at the build level, then that version takes precedence over this sourceVersion (at the project level).

For more information, see Source Version Sample with CodeBuild in the CodeBuild User Guide.

$sel:name:Project', project_name - The name of the build project.

$sel:vpcConfig:Project', project_vpcConfig - Information about the VPC configuration that CodeBuild accesses.

$sel:publicProjectAlias:Project', project_publicProjectAlias - Contains the project identifier used with the public build APIs.

$sel:source:Project', project_source - Information about the build input source code for this build project.

$sel:badge:Project', project_badge - Information about the build badge for the build project.

$sel:logsConfig:Project', project_logsConfig - Information about logs for the build project. A project can create logs in CloudWatch Logs, an S3 bucket, or both.

$sel:fileSystemLocations:Project', project_fileSystemLocations - An array of ProjectFileSystemLocation objects for a CodeBuild build project. A ProjectFileSystemLocation object specifies the identifier, location, mountOptions, mountPoint, and type of a file system created using Amazon Elastic File System.

$sel:buildBatchConfig:Project', project_buildBatchConfig - A ProjectBuildBatchConfig object that defines the batch build options for the project.

$sel:encryptionKey:Project', project_encryptionKey - The Key Management Service customer master key (CMK) to be used for encrypting the build output artifacts.

You can use a cross-account KMS key to encrypt the build output artifacts if your service role has permission to that key.

You can specify either the Amazon Resource Name (ARN) of the CMK or, if available, the CMK's alias (using the format alias/<alias-name>). If you don't specify a value, CodeBuild uses the managed CMK for Amazon Simple Storage Service (Amazon S3).

$sel:lastModified:Project', project_lastModified - When the build project's settings were last modified, expressed in Unix time format.

$sel:projectVisibility:Project', project_projectVisibility - Undocumented member.

$sel:webhook:Project', project_webhook - Information about a webhook that connects repository events to a build project in CodeBuild.

$sel:description:Project', project_description - A description that makes the build project easy to identify.

$sel:serviceRole:Project', project_serviceRole - The ARN of the IAM role that enables CodeBuild to interact with dependent Amazon Web Services services on behalf of the Amazon Web Services account.

$sel:tags:Project', project_tags - A list of tag key and value pairs associated with this build project.

These tags are available for use by Amazon Web Services services that support CodeBuild build project tags.

$sel:timeoutInMinutes:Project', project_timeoutInMinutes - How long, in minutes, from 5 to 480 (8 hours), for CodeBuild to wait before timing out any related build that did not get marked as completed. The default is 60 minutes.

project_secondaryArtifacts :: Lens' Project (Maybe [ProjectArtifacts])

An array of ProjectArtifacts objects.

project_resourceAccessRole :: Lens' Project (Maybe Text)

The ARN of the IAM role that enables CodeBuild to access the CloudWatch Logs and Amazon S3 artifacts for the project's builds.

project_arn :: Lens' Project (Maybe Text)

The Amazon Resource Name (ARN) of the build project.

project_artifacts :: Lens' Project (Maybe ProjectArtifacts)

Information about the build output artifacts for the build project.

project_environment :: Lens' Project (Maybe ProjectEnvironment)

Information about the build environment for this build project.

project_created :: Lens' Project (Maybe UTCTime)

When the build project was created, expressed in Unix time format.

project_concurrentBuildLimit :: Lens' Project (Maybe Int)

The maximum number of concurrent builds that are allowed for this project.

New builds are only started if the current number of builds is less than or equal to this limit. If the current build count meets this limit, new builds are throttled and are not run.

project_secondarySourceVersions :: Lens' Project (Maybe [ProjectSourceVersion])

An array of ProjectSourceVersion objects. If secondarySourceVersions is specified at the build level, then they take over these secondarySourceVersions (at the project level).

project_queuedTimeoutInMinutes :: Lens' Project (Maybe Natural)

The number of minutes a build is allowed to be queued before it times out.

project_cache :: Lens' Project (Maybe ProjectCache)

Information about the cache for the build project.

project_secondarySources :: Lens' Project (Maybe [ProjectSource])

An array of ProjectSource objects.

project_sourceVersion :: Lens' Project (Maybe Text)

A version of the build input to be built for this project. If not specified, the latest version is used. If specified, it must be one of:

For CodeCommit: the commit ID, branch, or Git tag to use.

For GitHub: the commit ID, pull request ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a pull request ID is specified, it must use the format pr/pull-request-ID (for example pr/25). If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

For Bitbucket: the commit ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

For Amazon S3: the version ID of the object that represents the build input ZIP file to use.

If sourceVersion is specified at the build level, then that version takes precedence over this sourceVersion (at the project level).

For more information, see Source Version Sample with CodeBuild in the CodeBuild User Guide.

project_name :: Lens' Project (Maybe Text)

The name of the build project.

project_vpcConfig :: Lens' Project (Maybe VpcConfig)

Information about the VPC configuration that CodeBuild accesses.

project_publicProjectAlias :: Lens' Project (Maybe Text)

Contains the project identifier used with the public build APIs.

project_source :: Lens' Project (Maybe ProjectSource)

Information about the build input source code for this build project.

project_badge :: Lens' Project (Maybe ProjectBadge)

Information about the build badge for the build project.

project_logsConfig :: Lens' Project (Maybe LogsConfig)

Information about logs for the build project. A project can create logs in CloudWatch Logs, an S3 bucket, or both.

project_fileSystemLocations :: Lens' Project (Maybe [ProjectFileSystemLocation])

An array of ProjectFileSystemLocation objects for a CodeBuild build project. A ProjectFileSystemLocation object specifies the identifier, location, mountOptions, mountPoint, and type of a file system created using Amazon Elastic File System.

project_buildBatchConfig :: Lens' Project (Maybe ProjectBuildBatchConfig)

A ProjectBuildBatchConfig object that defines the batch build options for the project.

project_encryptionKey :: Lens' Project (Maybe Text)

The Key Management Service customer master key (CMK) to be used for encrypting the build output artifacts.

You can use a cross-account KMS key to encrypt the build output artifacts if your service role has permission to that key.

You can specify either the Amazon Resource Name (ARN) of the CMK or, if available, the CMK's alias (using the format alias/<alias-name>). If you don't specify a value, CodeBuild uses the managed CMK for Amazon Simple Storage Service (Amazon S3).

project_lastModified :: Lens' Project (Maybe UTCTime)

When the build project's settings were last modified, expressed in Unix time format.

project_projectVisibility :: Lens' Project (Maybe ProjectVisibilityType)

Undocumented member.

project_webhook :: Lens' Project (Maybe Webhook)

Information about a webhook that connects repository events to a build project in CodeBuild.

project_description :: Lens' Project (Maybe Text)

A description that makes the build project easy to identify.

project_serviceRole :: Lens' Project (Maybe Text)

The ARN of the IAM role that enables CodeBuild to interact with dependent Amazon Web Services services on behalf of the Amazon Web Services account.

project_tags :: Lens' Project (Maybe [Tag])

A list of tag key and value pairs associated with this build project.

These tags are available for use by Amazon Web Services services that support CodeBuild build project tags.

project_timeoutInMinutes :: Lens' Project (Maybe Natural)

How long, in minutes, from 5 to 480 (8 hours), for CodeBuild to wait before timing out any related build that did not get marked as completed. The default is 60 minutes.