Network.AWS.CodeBuild.Types.ProjectArtifacts

Description

Documentation

data ProjectArtifacts

Information about the build output artifacts for the build project.

See: newProjectArtifacts smart constructor.

Constructors

ProjectArtifacts'

Fields

bucketOwnerAccess :: Maybe BucketOwnerAccess

namespaceType :: Maybe ArtifactNamespace

Along with path and name, the pattern that CodeBuild uses to determine the name and location to store the output artifact:

If type is set to CODEPIPELINE, CodePipeline ignores this value if specified. This is because CodePipeline manages its build output names instead of CodeBuild.

If type is set to NO_ARTIFACTS, this value is ignored if specified, because no build output is produced.

If type is set to S3, valid values include:

BUILD_ID: Include the build ID in the location of the build output artifact.

NONE: Do not include the build ID. This is the default if namespaceType is not specified.

For example, if path is set to MyArtifacts, namespaceType is set to BUILD_ID, and name is set to MyArtifact.zip, the output artifact is stored in MyArtifacts/<build-ID>/MyArtifact.zip.

artifactIdentifier :: Maybe Text

An identifier for this artifact definition.

overrideArtifactName :: Maybe Bool

If this flag is set, a name specified in the buildspec file overrides the artifact name. The name specified in a buildspec file is calculated at build time and uses the Shell Command Language. For example, you can append a date and time to your artifact name so that it is always unique.

encryptionDisabled :: Maybe Bool

Set to true if you do not want your output artifacts encrypted. This option is valid only if your artifacts type is Amazon S3. If this is set with another artifacts type, an invalidInputException is thrown.

name :: Maybe Text

Along with path and namespaceType, the pattern that CodeBuild uses to name and store the output artifact:

If type is set to CODEPIPELINE, CodePipeline ignores this value if specified. This is because CodePipeline manages its build output names instead of CodeBuild.

If type is set to NO_ARTIFACTS, this value is ignored if specified, because no build output is produced.

If type is set to S3, this is the name of the output artifact object. If you set the name to be a forward slash ("/"), the artifact is stored in the root of the output bucket.

For example:

If path is set to MyArtifacts, namespaceType is set to BUILD_ID, and name is set to MyArtifact.zip, then the output artifact is stored in MyArtifacts/<build-ID>/MyArtifact.zip.

If path is empty, namespaceType is set to NONE, and name is set to "/", the output artifact is stored in the root of the output bucket.

If path is set to MyArtifacts, namespaceType is set to BUILD_ID, and name is set to "/", the output artifact is stored in MyArtifacts/<build-ID>.

packaging :: Maybe ArtifactPackaging

The type of build output artifact to create:

If type is set to CODEPIPELINE, CodePipeline ignores this value if specified. This is because CodePipeline manages its build output artifacts instead of CodeBuild.

If type is set to NO_ARTIFACTS, this value is ignored if specified, because no build output is produced.

If type is set to S3, valid values include:

NONE: CodeBuild creates in the output bucket a folder that contains the build output. This is the default if packaging is not specified.

ZIP: CodeBuild creates in the output bucket a ZIP file that contains the build output.

path :: Maybe Text

Along with namespaceType and name, the pattern that CodeBuild uses to name and store the output artifact:

If type is set to CODEPIPELINE, CodePipeline ignores this value if specified. This is because CodePipeline manages its build output names instead of CodeBuild.

If type is set to NO_ARTIFACTS, this value is ignored if specified, because no build output is produced.

If type is set to S3, this is the path to the output artifact. If path is not specified, path is not used.

For example, if path is set to MyArtifacts, namespaceType is set to NONE, and name is set to MyArtifact.zip, the output artifact is stored in the output bucket at MyArtifacts/MyArtifact.zip.

location :: Maybe Text

Information about the build output artifact location:

If type is set to CODEPIPELINE, CodePipeline ignores this value if specified. This is because CodePipeline manages its build output locations instead of CodeBuild.

If type is set to NO_ARTIFACTS, this value is ignored if specified, because no build output is produced.

If type is set to S3, this is the name of the output bucket.

type' :: ArtifactsType

The type of build output artifact. Valid values include:

CODEPIPELINE: The build project has build output generated through CodePipeline.

The CODEPIPELINE type is not supported for secondaryArtifacts.

NO_ARTIFACTS: The build project does not produce any build output.

S3: The build project stores build output in Amazon S3.

newProjectArtifacts

Arguments

:: ArtifactsType
$sel:type':ProjectArtifacts'

-> ProjectArtifacts

Create a value of ProjectArtifacts with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bucketOwnerAccess:ProjectArtifacts', projectArtifacts_bucketOwnerAccess - Undocumented member.

$sel:namespaceType:ProjectArtifacts', projectArtifacts_namespaceType - Along with path and name, the pattern that CodeBuild uses to determine the name and location to store the output artifact:

If type is set to CODEPIPELINE, CodePipeline ignores this value if specified. This is because CodePipeline manages its build output names instead of CodeBuild.

If type is set to NO_ARTIFACTS, this value is ignored if specified, because no build output is produced.

If type is set to S3, valid values include:

BUILD_ID: Include the build ID in the location of the build output artifact.

NONE: Do not include the build ID. This is the default if namespaceType is not specified.

For example, if path is set to MyArtifacts, namespaceType is set to BUILD_ID, and name is set to MyArtifact.zip, the output artifact is stored in MyArtifacts/<build-ID>/MyArtifact.zip.

$sel:artifactIdentifier:ProjectArtifacts', projectArtifacts_artifactIdentifier - An identifier for this artifact definition.

$sel:overrideArtifactName:ProjectArtifacts', projectArtifacts_overrideArtifactName - If this flag is set, a name specified in the buildspec file overrides the artifact name. The name specified in a buildspec file is calculated at build time and uses the Shell Command Language. For example, you can append a date and time to your artifact name so that it is always unique.

$sel:encryptionDisabled:ProjectArtifacts', projectArtifacts_encryptionDisabled - Set to true if you do not want your output artifacts encrypted. This option is valid only if your artifacts type is Amazon S3. If this is set with another artifacts type, an invalidInputException is thrown.

$sel:name:ProjectArtifacts', projectArtifacts_name - Along with path and namespaceType, the pattern that CodeBuild uses to name and store the output artifact:

If type is set to CODEPIPELINE, CodePipeline ignores this value if specified. This is because CodePipeline manages its build output names instead of CodeBuild.

If type is set to NO_ARTIFACTS, this value is ignored if specified, because no build output is produced.

If type is set to S3, this is the name of the output artifact object. If you set the name to be a forward slash ("/"), the artifact is stored in the root of the output bucket.

For example:

If path is set to MyArtifacts, namespaceType is set to BUILD_ID, and name is set to MyArtifact.zip, then the output artifact is stored in MyArtifacts/<build-ID>/MyArtifact.zip.

If path is empty, namespaceType is set to NONE, and name is set to "/", the output artifact is stored in the root of the output bucket.

If path is set to MyArtifacts, namespaceType is set to BUILD_ID, and name is set to "/", the output artifact is stored in MyArtifacts/<build-ID>.

$sel:packaging:ProjectArtifacts', projectArtifacts_packaging - The type of build output artifact to create:

If type is set to CODEPIPELINE, CodePipeline ignores this value if specified. This is because CodePipeline manages its build output artifacts instead of CodeBuild.

If type is set to NO_ARTIFACTS, this value is ignored if specified, because no build output is produced.

If type is set to S3, valid values include:

NONE: CodeBuild creates in the output bucket a folder that contains the build output. This is the default if packaging is not specified.

ZIP: CodeBuild creates in the output bucket a ZIP file that contains the build output.

$sel:path:ProjectArtifacts', projectArtifacts_path - Along with namespaceType and name, the pattern that CodeBuild uses to name and store the output artifact:

If type is set to CODEPIPELINE, CodePipeline ignores this value if specified. This is because CodePipeline manages its build output names instead of CodeBuild.

If type is set to NO_ARTIFACTS, this value is ignored if specified, because no build output is produced.

If type is set to S3, this is the path to the output artifact. If path is not specified, path is not used.

For example, if path is set to MyArtifacts, namespaceType is set to NONE, and name is set to MyArtifact.zip, the output artifact is stored in the output bucket at MyArtifacts/MyArtifact.zip.

$sel:location:ProjectArtifacts', projectArtifacts_location - Information about the build output artifact location:

If type is set to CODEPIPELINE, CodePipeline ignores this value if specified. This is because CodePipeline manages its build output locations instead of CodeBuild.

If type is set to NO_ARTIFACTS, this value is ignored if specified, because no build output is produced.

If type is set to S3, this is the name of the output bucket.

$sel:type':ProjectArtifacts', projectArtifacts_type - The type of build output artifact. Valid values include:

CODEPIPELINE: The build project has build output generated through CodePipeline.

The CODEPIPELINE type is not supported for secondaryArtifacts.

NO_ARTIFACTS: The build project does not produce any build output.

S3: The build project stores build output in Amazon S3.

projectArtifacts_bucketOwnerAccess :: Lens' ProjectArtifacts (Maybe BucketOwnerAccess)

Undocumented member.

projectArtifacts_namespaceType :: Lens' ProjectArtifacts (Maybe ArtifactNamespace)

Along with path and name, the pattern that CodeBuild uses to determine the name and location to store the output artifact:

If type is set to CODEPIPELINE, CodePipeline ignores this value if specified. This is because CodePipeline manages its build output names instead of CodeBuild.

If type is set to NO_ARTIFACTS, this value is ignored if specified, because no build output is produced.

If type is set to S3, valid values include:

BUILD_ID: Include the build ID in the location of the build output artifact.

NONE: Do not include the build ID. This is the default if namespaceType is not specified.

For example, if path is set to MyArtifacts, namespaceType is set to BUILD_ID, and name is set to MyArtifact.zip, the output artifact is stored in MyArtifacts/<build-ID>/MyArtifact.zip.

projectArtifacts_artifactIdentifier :: Lens' ProjectArtifacts (Maybe Text)

An identifier for this artifact definition.

projectArtifacts_overrideArtifactName :: Lens' ProjectArtifacts (Maybe Bool)

If this flag is set, a name specified in the buildspec file overrides the artifact name. The name specified in a buildspec file is calculated at build time and uses the Shell Command Language. For example, you can append a date and time to your artifact name so that it is always unique.

projectArtifacts_encryptionDisabled :: Lens' ProjectArtifacts (Maybe Bool)

Set to true if you do not want your output artifacts encrypted. This option is valid only if your artifacts type is Amazon S3. If this is set with another artifacts type, an invalidInputException is thrown.

projectArtifacts_name :: Lens' ProjectArtifacts (Maybe Text)

Along with path and namespaceType, the pattern that CodeBuild uses to name and store the output artifact:

If type is set to CODEPIPELINE, CodePipeline ignores this value if specified. This is because CodePipeline manages its build output names instead of CodeBuild.

If type is set to NO_ARTIFACTS, this value is ignored if specified, because no build output is produced.

If type is set to S3, this is the name of the output artifact object. If you set the name to be a forward slash ("/"), the artifact is stored in the root of the output bucket.

For example:

If path is set to MyArtifacts, namespaceType is set to BUILD_ID, and name is set to MyArtifact.zip, then the output artifact is stored in MyArtifacts/<build-ID>/MyArtifact.zip.

If path is empty, namespaceType is set to NONE, and name is set to "/", the output artifact is stored in the root of the output bucket.

If path is set to MyArtifacts, namespaceType is set to BUILD_ID, and name is set to "/", the output artifact is stored in MyArtifacts/<build-ID>.

projectArtifacts_packaging :: Lens' ProjectArtifacts (Maybe ArtifactPackaging)

The type of build output artifact to create:

If type is set to CODEPIPELINE, CodePipeline ignores this value if specified. This is because CodePipeline manages its build output artifacts instead of CodeBuild.

If type is set to NO_ARTIFACTS, this value is ignored if specified, because no build output is produced.

If type is set to S3, valid values include:

NONE: CodeBuild creates in the output bucket a folder that contains the build output. This is the default if packaging is not specified.

ZIP: CodeBuild creates in the output bucket a ZIP file that contains the build output.

projectArtifacts_path :: Lens' ProjectArtifacts (Maybe Text)

Along with namespaceType and name, the pattern that CodeBuild uses to name and store the output artifact:

If type is set to CODEPIPELINE, CodePipeline ignores this value if specified. This is because CodePipeline manages its build output names instead of CodeBuild.

If type is set to NO_ARTIFACTS, this value is ignored if specified, because no build output is produced.

If type is set to S3, this is the path to the output artifact. If path is not specified, path is not used.

For example, if path is set to MyArtifacts, namespaceType is set to NONE, and name is set to MyArtifact.zip, the output artifact is stored in the output bucket at MyArtifacts/MyArtifact.zip.

projectArtifacts_location :: Lens' ProjectArtifacts (Maybe Text)

Information about the build output artifact location:

If type is set to CODEPIPELINE, CodePipeline ignores this value if specified. This is because CodePipeline manages its build output locations instead of CodeBuild.

If type is set to NO_ARTIFACTS, this value is ignored if specified, because no build output is produced.

If type is set to S3, this is the name of the output bucket.

projectArtifacts_type :: Lens' ProjectArtifacts ArtifactsType

The type of build output artifact. Valid values include:

CODEPIPELINE: The build project has build output generated through CodePipeline.

The CODEPIPELINE type is not supported for secondaryArtifacts.

NO_ARTIFACTS: The build project does not produce any build output.

S3: The build project stores build output in Amazon S3.