Network.AWS.CodeBuild.Types.ProjectBuildBatchConfig

Description

Documentation

data ProjectBuildBatchConfig

Contains configuration information about a batch build project.

See: newProjectBuildBatchConfig smart constructor.

Constructors

ProjectBuildBatchConfig'

Fields

combineArtifacts :: Maybe Bool

Specifies if the build artifacts for the batch build should be combined into a single artifact location.

timeoutInMins :: Maybe Int

Specifies the maximum amount of time, in minutes, that the batch build must be completed in.

restrictions :: Maybe BatchRestrictions

A BatchRestrictions object that specifies the restrictions for the batch build.

batchReportMode :: Maybe BatchReportModeType

Specifies how build status reports are sent to the source provider for the batch build. This property is only used when the source provider for your project is Bitbucket, GitHub, or GitHub Enterprise, and your project is configured to report build statuses to the source provider.

REPORT_AGGREGATED_BATCH Aggregate all of the build statuses into a single status report.

REPORT_INDIVIDUAL_BUILDS

Send a separate status report for each individual build.

serviceRole :: Maybe Text

Specifies the service role ARN for the batch build project.

newProjectBuildBatchConfig :: ProjectBuildBatchConfig

Create a value of ProjectBuildBatchConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:combineArtifacts:ProjectBuildBatchConfig', projectBuildBatchConfig_combineArtifacts - Specifies if the build artifacts for the batch build should be combined into a single artifact location.

$sel:timeoutInMins:ProjectBuildBatchConfig', projectBuildBatchConfig_timeoutInMins - Specifies the maximum amount of time, in minutes, that the batch build must be completed in.

$sel:restrictions:ProjectBuildBatchConfig', projectBuildBatchConfig_restrictions - A BatchRestrictions object that specifies the restrictions for the batch build.

$sel:batchReportMode:ProjectBuildBatchConfig', projectBuildBatchConfig_batchReportMode - Specifies how build status reports are sent to the source provider for the batch build. This property is only used when the source provider for your project is Bitbucket, GitHub, or GitHub Enterprise, and your project is configured to report build statuses to the source provider.

REPORT_AGGREGATED_BATCH Aggregate all of the build statuses into a single status report.

REPORT_INDIVIDUAL_BUILDS

Send a separate status report for each individual build.

$sel:serviceRole:ProjectBuildBatchConfig', projectBuildBatchConfig_serviceRole - Specifies the service role ARN for the batch build project.

projectBuildBatchConfig_combineArtifacts :: Lens' ProjectBuildBatchConfig (Maybe Bool)

Specifies if the build artifacts for the batch build should be combined into a single artifact location.

projectBuildBatchConfig_timeoutInMins :: Lens' ProjectBuildBatchConfig (Maybe Int)

Specifies the maximum amount of time, in minutes, that the batch build must be completed in.

projectBuildBatchConfig_restrictions :: Lens' ProjectBuildBatchConfig (Maybe BatchRestrictions)

A BatchRestrictions object that specifies the restrictions for the batch build.

projectBuildBatchConfig_batchReportMode :: Lens' ProjectBuildBatchConfig (Maybe BatchReportModeType)

Specifies how build status reports are sent to the source provider for the batch build. This property is only used when the source provider for your project is Bitbucket, GitHub, or GitHub Enterprise, and your project is configured to report build statuses to the source provider.

REPORT_AGGREGATED_BATCH Aggregate all of the build statuses into a single status report.

REPORT_INDIVIDUAL_BUILDS

Send a separate status report for each individual build.

projectBuildBatchConfig_serviceRole :: Lens' ProjectBuildBatchConfig (Maybe Text)

Specifies the service role ARN for the batch build project.