Network.AWS.CodeBuild.Types.ProjectSource

Description

Documentation

data ProjectSource

Information about the build input source code for the build project.

See: newProjectSource smart constructor.

Constructors

ProjectSource'

Fields

reportBuildStatus :: Maybe Bool

Set to true to report the status of a build's start and finish to your source provider. This option is valid only when your source provider is GitHub, GitHub Enterprise, or Bitbucket. If this is set and you use a different source provider, an invalidInputException is thrown.

To be able to report the build status to the source provider, the user associated with the source provider must have write access to the repo. If the user does not have write access, the build status cannot be updated. For more information, see Source provider access in the CodeBuild User Guide.

The status of a build triggered by a webhook is always reported to your source provider.

If your project's builds are triggered by a webhook, you must push a new commit to the repo for a change to this property to take effect.

insecureSsl :: Maybe Bool

Enable this flag to ignore SSL warnings while connecting to the project source code.

location :: Maybe Text

Information about the location of the source code to be built. Valid values include:

For source code settings that are specified in the source action of a pipeline in CodePipeline, location should not be specified. If it is specified, CodePipeline ignores it. This is because CodePipeline uses the settings in a pipeline's source action instead of this value.

For source code in an CodeCommit repository, the HTTPS clone URL to the repository that contains the source code and the buildspec file (for example, https://git-codecommit.<region-ID>.amazonaws.com/v1/repos/<repo-name>).

For source code in an Amazon S3 input bucket, one of the following.

The path to the ZIP file that contains the source code (for example, <bucket-name>/<path>/<object-name>.zip).

The path to the folder that contains the source code (for example, <bucket-name>/<path-to-source-code>/<folder>/).

For source code in a GitHub repository, the HTTPS clone URL to the repository that contains the source and the buildspec file. You must connect your Amazon Web Services account to your GitHub account. Use the CodeBuild console to start creating a build project. When you use the console to connect (or reconnect) with GitHub, on the GitHub Authorize application page, for Organization access, choose Request access next to each repository you want to allow CodeBuild to have access to, and then choose __Authorize application__. (After you have connected to your GitHub account, you do not need to finish creating the build project. You can leave the CodeBuild console.) To instruct CodeBuild to use this connection, in the source object, set the auth object's type value to OAUTH.

For source code in a Bitbucket repository, the HTTPS clone URL to the repository that contains the source and the buildspec file. You must connect your Amazon Web Services account to your Bitbucket account. Use the CodeBuild console to start creating a build project. When you use the console to connect (or reconnect) with Bitbucket, on the Bitbucket Confirm access to your account page, choose Grant access. (After you have connected to your Bitbucket account, you do not need to finish creating the build project. You can leave the CodeBuild console.) To instruct CodeBuild to use this connection, in the source object, set the auth object's type value to OAUTH.

If you specify CODEPIPELINE for the Type property, don't specify this property. For all of the other types, you must specify Location.

auth :: Maybe SourceAuth

Information about the authorization settings for CodeBuild to access the source code to be built.

This information is for the CodeBuild console's use only. Your code should not get or set this information directly.

buildspec :: Maybe Text

The buildspec file declaration to use for the builds in this build project.

If this value is set, it can be either an inline buildspec definition, the path to an alternate buildspec file relative to the value of the built-in CODEBUILD_SRC_DIR environment variable, or the path to an S3 bucket. The bucket must be in the same Amazon Web Services Region as the build project. Specify the buildspec file using its ARN (for example, arn:aws:s3:::my-codebuild-sample2/buildspec.yml). If this value is not provided or is set to an empty string, the source code must contain a buildspec file in its root directory. For more information, see Buildspec File Name and Storage Location.

sourceIdentifier :: Maybe Text

An identifier for this project source. The identifier can only contain alphanumeric characters and underscores, and must be less than 128 characters in length.

gitCloneDepth :: Maybe Natural

Information about the Git clone depth for the build project.

gitSubmodulesConfig :: Maybe GitSubmodulesConfig

Information about the Git submodules configuration for the build project.

buildStatusConfig :: Maybe BuildStatusConfig

Contains information that defines how the build project reports the build status to the source provider. This option is only used when the source provider is GITHUB, GITHUB_ENTERPRISE, or BITBUCKET.

type' :: SourceType

The type of repository that contains the source code to be built. Valid values include:

BITBUCKET: The source code is in a Bitbucket repository.

CODECOMMIT: The source code is in an CodeCommit repository.

CODEPIPELINE: The source code settings are specified in the source action of a pipeline in CodePipeline.

GITHUB: The source code is in a GitHub or GitHub Enterprise Cloud repository.

GITHUB_ENTERPRISE: The source code is in a GitHub Enterprise Server repository.

NO_SOURCE: The project does not have input source code.

S3: The source code is in an Amazon S3 bucket.

newProjectSource

Arguments

:: SourceType
$sel:type':ProjectSource'

-> ProjectSource

Create a value of ProjectSource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reportBuildStatus:ProjectSource', projectSource_reportBuildStatus - Set to true to report the status of a build's start and finish to your source provider. This option is valid only when your source provider is GitHub, GitHub Enterprise, or Bitbucket. If this is set and you use a different source provider, an invalidInputException is thrown.

To be able to report the build status to the source provider, the user associated with the source provider must have write access to the repo. If the user does not have write access, the build status cannot be updated. For more information, see Source provider access in the CodeBuild User Guide.

The status of a build triggered by a webhook is always reported to your source provider.

If your project's builds are triggered by a webhook, you must push a new commit to the repo for a change to this property to take effect.

$sel:insecureSsl:ProjectSource', projectSource_insecureSsl - Enable this flag to ignore SSL warnings while connecting to the project source code.

$sel:location:ProjectSource', projectSource_location - Information about the location of the source code to be built. Valid values include:

For source code settings that are specified in the source action of a pipeline in CodePipeline, location should not be specified. If it is specified, CodePipeline ignores it. This is because CodePipeline uses the settings in a pipeline's source action instead of this value.

For source code in an CodeCommit repository, the HTTPS clone URL to the repository that contains the source code and the buildspec file (for example, https://git-codecommit.<region-ID>.amazonaws.com/v1/repos/<repo-name>).

For source code in an Amazon S3 input bucket, one of the following.

The path to the ZIP file that contains the source code (for example, <bucket-name>/<path>/<object-name>.zip).

The path to the folder that contains the source code (for example, <bucket-name>/<path-to-source-code>/<folder>/).

For source code in a GitHub repository, the HTTPS clone URL to the repository that contains the source and the buildspec file. You must connect your Amazon Web Services account to your GitHub account. Use the CodeBuild console to start creating a build project. When you use the console to connect (or reconnect) with GitHub, on the GitHub Authorize application page, for Organization access, choose Request access next to each repository you want to allow CodeBuild to have access to, and then choose __Authorize application__. (After you have connected to your GitHub account, you do not need to finish creating the build project. You can leave the CodeBuild console.) To instruct CodeBuild to use this connection, in the source object, set the auth object's type value to OAUTH.

For source code in a Bitbucket repository, the HTTPS clone URL to the repository that contains the source and the buildspec file. You must connect your Amazon Web Services account to your Bitbucket account. Use the CodeBuild console to start creating a build project. When you use the console to connect (or reconnect) with Bitbucket, on the Bitbucket Confirm access to your account page, choose Grant access. (After you have connected to your Bitbucket account, you do not need to finish creating the build project. You can leave the CodeBuild console.) To instruct CodeBuild to use this connection, in the source object, set the auth object's type value to OAUTH.

If you specify CODEPIPELINE for the Type property, don't specify this property. For all of the other types, you must specify Location.

$sel:auth:ProjectSource', projectSource_auth - Information about the authorization settings for CodeBuild to access the source code to be built.

This information is for the CodeBuild console's use only. Your code should not get or set this information directly.

$sel:buildspec:ProjectSource', projectSource_buildspec - The buildspec file declaration to use for the builds in this build project.

If this value is set, it can be either an inline buildspec definition, the path to an alternate buildspec file relative to the value of the built-in CODEBUILD_SRC_DIR environment variable, or the path to an S3 bucket. The bucket must be in the same Amazon Web Services Region as the build project. Specify the buildspec file using its ARN (for example, arn:aws:s3:::my-codebuild-sample2/buildspec.yml). If this value is not provided or is set to an empty string, the source code must contain a buildspec file in its root directory. For more information, see Buildspec File Name and Storage Location.

$sel:sourceIdentifier:ProjectSource', projectSource_sourceIdentifier - An identifier for this project source. The identifier can only contain alphanumeric characters and underscores, and must be less than 128 characters in length.

$sel:gitCloneDepth:ProjectSource', projectSource_gitCloneDepth - Information about the Git clone depth for the build project.

$sel:gitSubmodulesConfig:ProjectSource', projectSource_gitSubmodulesConfig - Information about the Git submodules configuration for the build project.

$sel:buildStatusConfig:ProjectSource', projectSource_buildStatusConfig - Contains information that defines how the build project reports the build status to the source provider. This option is only used when the source provider is GITHUB, GITHUB_ENTERPRISE, or BITBUCKET.

$sel:type':ProjectSource', projectSource_type - The type of repository that contains the source code to be built. Valid values include:

BITBUCKET: The source code is in a Bitbucket repository.

CODECOMMIT: The source code is in an CodeCommit repository.

CODEPIPELINE: The source code settings are specified in the source action of a pipeline in CodePipeline.

GITHUB: The source code is in a GitHub or GitHub Enterprise Cloud repository.

GITHUB_ENTERPRISE: The source code is in a GitHub Enterprise Server repository.

NO_SOURCE: The project does not have input source code.

S3: The source code is in an Amazon S3 bucket.

projectSource_reportBuildStatus :: Lens' ProjectSource (Maybe Bool)

Set to true to report the status of a build's start and finish to your source provider. This option is valid only when your source provider is GitHub, GitHub Enterprise, or Bitbucket. If this is set and you use a different source provider, an invalidInputException is thrown.

To be able to report the build status to the source provider, the user associated with the source provider must have write access to the repo. If the user does not have write access, the build status cannot be updated. For more information, see Source provider access in the CodeBuild User Guide.

The status of a build triggered by a webhook is always reported to your source provider.

If your project's builds are triggered by a webhook, you must push a new commit to the repo for a change to this property to take effect.

projectSource_insecureSsl :: Lens' ProjectSource (Maybe Bool)

Enable this flag to ignore SSL warnings while connecting to the project source code.

projectSource_location :: Lens' ProjectSource (Maybe Text)

Information about the location of the source code to be built. Valid values include:

For source code settings that are specified in the source action of a pipeline in CodePipeline, location should not be specified. If it is specified, CodePipeline ignores it. This is because CodePipeline uses the settings in a pipeline's source action instead of this value.

For source code in an CodeCommit repository, the HTTPS clone URL to the repository that contains the source code and the buildspec file (for example, https://git-codecommit.<region-ID>.amazonaws.com/v1/repos/<repo-name>).

For source code in an Amazon S3 input bucket, one of the following.

The path to the ZIP file that contains the source code (for example, <bucket-name>/<path>/<object-name>.zip).

The path to the folder that contains the source code (for example, <bucket-name>/<path-to-source-code>/<folder>/).

For source code in a GitHub repository, the HTTPS clone URL to the repository that contains the source and the buildspec file. You must connect your Amazon Web Services account to your GitHub account. Use the CodeBuild console to start creating a build project. When you use the console to connect (or reconnect) with GitHub, on the GitHub Authorize application page, for Organization access, choose Request access next to each repository you want to allow CodeBuild to have access to, and then choose __Authorize application__. (After you have connected to your GitHub account, you do not need to finish creating the build project. You can leave the CodeBuild console.) To instruct CodeBuild to use this connection, in the source object, set the auth object's type value to OAUTH.

For source code in a Bitbucket repository, the HTTPS clone URL to the repository that contains the source and the buildspec file. You must connect your Amazon Web Services account to your Bitbucket account. Use the CodeBuild console to start creating a build project. When you use the console to connect (or reconnect) with Bitbucket, on the Bitbucket Confirm access to your account page, choose Grant access. (After you have connected to your Bitbucket account, you do not need to finish creating the build project. You can leave the CodeBuild console.) To instruct CodeBuild to use this connection, in the source object, set the auth object's type value to OAUTH.

If you specify CODEPIPELINE for the Type property, don't specify this property. For all of the other types, you must specify Location.

projectSource_auth :: Lens' ProjectSource (Maybe SourceAuth)

Information about the authorization settings for CodeBuild to access the source code to be built.

This information is for the CodeBuild console's use only. Your code should not get or set this information directly.

projectSource_buildspec :: Lens' ProjectSource (Maybe Text)

The buildspec file declaration to use for the builds in this build project.

If this value is set, it can be either an inline buildspec definition, the path to an alternate buildspec file relative to the value of the built-in CODEBUILD_SRC_DIR environment variable, or the path to an S3 bucket. The bucket must be in the same Amazon Web Services Region as the build project. Specify the buildspec file using its ARN (for example, arn:aws:s3:::my-codebuild-sample2/buildspec.yml). If this value is not provided or is set to an empty string, the source code must contain a buildspec file in its root directory. For more information, see Buildspec File Name and Storage Location.

projectSource_sourceIdentifier :: Lens' ProjectSource (Maybe Text)

An identifier for this project source. The identifier can only contain alphanumeric characters and underscores, and must be less than 128 characters in length.

projectSource_gitCloneDepth :: Lens' ProjectSource (Maybe Natural)

Information about the Git clone depth for the build project.

projectSource_gitSubmodulesConfig :: Lens' ProjectSource (Maybe GitSubmodulesConfig)

Information about the Git submodules configuration for the build project.

projectSource_buildStatusConfig :: Lens' ProjectSource (Maybe BuildStatusConfig)

Contains information that defines how the build project reports the build status to the source provider. This option is only used when the source provider is GITHUB, GITHUB_ENTERPRISE, or BITBUCKET.

projectSource_type :: Lens' ProjectSource SourceType

The type of repository that contains the source code to be built. Valid values include:

BITBUCKET: The source code is in a Bitbucket repository.

CODECOMMIT: The source code is in an CodeCommit repository.

CODEPIPELINE: The source code settings are specified in the source action of a pipeline in CodePipeline.

GITHUB: The source code is in a GitHub or GitHub Enterprise Cloud repository.

GITHUB_ENTERPRISE: The source code is in a GitHub Enterprise Server repository.

NO_SOURCE: The project does not have input source code.

S3: The source code is in an Amazon S3 bucket.