Network.AWS.CodeBuild.Types.Report

Description

Documentation

data Report

Information about the results from running a series of test cases during the run of a build project. The test cases are specified in the buildspec for the build project using one or more paths to the test case files. You can specify any type of tests you want, such as unit tests, integration tests, and functional tests.

See: newReport smart constructor.

Constructors

Report'

Fields

codeCoverageSummary :: Maybe CodeCoverageReportSummary

A CodeCoverageReportSummary object that contains a code coverage summary for this report.

status :: Maybe ReportStatusType

The status of this report.

reportGroupArn :: Maybe Text

The ARN of the report group associated with this report.

exportConfig :: Maybe ReportExportConfig

Information about where the raw data used to generate this report was exported.

arn :: Maybe Text

The ARN of the report run.

testSummary :: Maybe TestReportSummary

A TestReportSummary object that contains information about this test report.

executionId :: Maybe Text

The ARN of the build run that generated this report.

expired :: Maybe POSIX

The date and time a report expires. A report expires 30 days after it is created. An expired report is not available to view in CodeBuild.

name :: Maybe Text

The name of the report that was run.

type' :: Maybe ReportType

The type of the report that was run.

CODE_COVERAGE

A code coverage report.

TEST

A test report.

created :: Maybe POSIX

The date and time this report run occurred.

truncated :: Maybe Bool

A boolean that specifies if this report run is truncated. The list of test cases is truncated after the maximum number of test cases is reached.

newReport :: Report

Create a value of Report with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:codeCoverageSummary:Report', report_codeCoverageSummary - A CodeCoverageReportSummary object that contains a code coverage summary for this report.

$sel:status:Report', report_status - The status of this report.

$sel:reportGroupArn:Report', report_reportGroupArn - The ARN of the report group associated with this report.

$sel:exportConfig:Report', report_exportConfig - Information about where the raw data used to generate this report was exported.

$sel:arn:Report', report_arn - The ARN of the report run.

$sel:testSummary:Report', report_testSummary - A TestReportSummary object that contains information about this test report.

$sel:executionId:Report', report_executionId - The ARN of the build run that generated this report.

$sel:expired:Report', report_expired - The date and time a report expires. A report expires 30 days after it is created. An expired report is not available to view in CodeBuild.

$sel:name:Report', report_name - The name of the report that was run.

$sel:type':Report', report_type - The type of the report that was run.

CODE_COVERAGE

A code coverage report.

TEST

A test report.

$sel:created:Report', report_created - The date and time this report run occurred.

$sel:truncated:Report', report_truncated - A boolean that specifies if this report run is truncated. The list of test cases is truncated after the maximum number of test cases is reached.

report_codeCoverageSummary :: Lens' Report (Maybe CodeCoverageReportSummary)

A CodeCoverageReportSummary object that contains a code coverage summary for this report.

report_status :: Lens' Report (Maybe ReportStatusType)

The status of this report.

report_reportGroupArn :: Lens' Report (Maybe Text)

The ARN of the report group associated with this report.

report_exportConfig :: Lens' Report (Maybe ReportExportConfig)

Information about where the raw data used to generate this report was exported.

report_arn :: Lens' Report (Maybe Text)

The ARN of the report run.

report_testSummary :: Lens' Report (Maybe TestReportSummary)

A TestReportSummary object that contains information about this test report.

report_executionId :: Lens' Report (Maybe Text)

The ARN of the build run that generated this report.

report_expired :: Lens' Report (Maybe UTCTime)

The date and time a report expires. A report expires 30 days after it is created. An expired report is not available to view in CodeBuild.

report_name :: Lens' Report (Maybe Text)

The name of the report that was run.

report_type :: Lens' Report (Maybe ReportType)

The type of the report that was run.

CODE_COVERAGE

A code coverage report.

TEST

A test report.

report_created :: Lens' Report (Maybe UTCTime)

The date and time this report run occurred.

report_truncated :: Lens' Report (Maybe Bool)

A boolean that specifies if this report run is truncated. The list of test cases is truncated after the maximum number of test cases is reached.