Network.AWS.CodeBuild.Types.ReportFilter

Description

Documentation

data ReportFilter

A filter used to return reports with the status specified by the input status parameter.

See: newReportFilter smart constructor.

Constructors

ReportFilter'

Fields

status :: Maybe ReportStatusType

The status used to filter reports. You can filter using one status only.

newReportFilter :: ReportFilter

Create a value of ReportFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ReportFilter', reportFilter_status - The status used to filter reports. You can filter using one status only.

reportFilter_status :: Lens' ReportFilter (Maybe ReportStatusType)

The status used to filter reports. You can filter using one status only.