Network.AWS.CodeBuild.Types.TestReportSummary

Description

Documentation

data TestReportSummary

Information about a test report.

See: newTestReportSummary smart constructor.

Constructors

TestReportSummary'

Fields

total :: Int

The number of test cases in this TestReportSummary. The total includes truncated test cases.

statusCounts :: HashMap Text Int

A map that contains the number of each type of status returned by the test results in this TestReportSummary.

durationInNanoSeconds :: Integer

The number of nanoseconds it took to run all of the test cases in this report.

newTestReportSummary

Arguments

:: Int
$sel:total:TestReportSummary'

-> Integer
$sel:durationInNanoSeconds:TestReportSummary'

-> TestReportSummary

Create a value of TestReportSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:total:TestReportSummary', testReportSummary_total - The number of test cases in this TestReportSummary. The total includes truncated test cases.

$sel:statusCounts:TestReportSummary', testReportSummary_statusCounts - A map that contains the number of each type of status returned by the test results in this TestReportSummary.

$sel:durationInNanoSeconds:TestReportSummary', testReportSummary_durationInNanoSeconds - The number of nanoseconds it took to run all of the test cases in this report.

testReportSummary_total :: Lens' TestReportSummary Int

The number of test cases in this TestReportSummary. The total includes truncated test cases.

testReportSummary_statusCounts :: Lens' TestReportSummary (HashMap Text Int)

A map that contains the number of each type of status returned by the test results in this TestReportSummary.

testReportSummary_durationInNanoSeconds :: Lens' TestReportSummary Integer

The number of nanoseconds it took to run all of the test cases in this report.