Network.AWS.CodeBuild.Types.Webhook

Description

Documentation

data Webhook

Information about a webhook that connects repository events to a build project in CodeBuild.

See: newWebhook smart constructor.

Constructors

Webhook'

Fields

branchFilter :: Maybe Text

A regular expression used to determine which repository branches are built when a webhook is triggered. If the name of a branch matches the regular expression, then it is built. If branchFilter is empty, then all branches are built.

It is recommended that you use filterGroups instead of branchFilter.

lastModifiedSecret :: Maybe POSIX

A timestamp that indicates the last time a repository's secret token was modified.

url :: Maybe Text

The URL to the webhook.

secret :: Maybe Text

The secret token of the associated repository.

A Bitbucket webhook does not support secret.

filterGroups :: Maybe [[WebhookFilter]]

An array of arrays of WebhookFilter objects used to determine which webhooks are triggered. At least one WebhookFilter in the array must specify EVENT as its type.

For a build to be triggered, at least one filter group in the filterGroups array must pass. For a filter group to pass, each of its filters must pass.

payloadUrl :: Maybe Text

The CodeBuild endpoint where webhook events are sent.

buildType :: Maybe WebhookBuildType

Specifies the type of build this webhook will trigger.

newWebhook :: Webhook

Create a value of Webhook with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:branchFilter:Webhook', webhook_branchFilter - A regular expression used to determine which repository branches are built when a webhook is triggered. If the name of a branch matches the regular expression, then it is built. If branchFilter is empty, then all branches are built.

It is recommended that you use filterGroups instead of branchFilter.

$sel:lastModifiedSecret:Webhook', webhook_lastModifiedSecret - A timestamp that indicates the last time a repository's secret token was modified.

$sel:url:Webhook', webhook_url - The URL to the webhook.

$sel:secret:Webhook', webhook_secret - The secret token of the associated repository.

A Bitbucket webhook does not support secret.

$sel:filterGroups:Webhook', webhook_filterGroups - An array of arrays of WebhookFilter objects used to determine which webhooks are triggered. At least one WebhookFilter in the array must specify EVENT as its type.

For a build to be triggered, at least one filter group in the filterGroups array must pass. For a filter group to pass, each of its filters must pass.

$sel:payloadUrl:Webhook', webhook_payloadUrl - The CodeBuild endpoint where webhook events are sent.

$sel:buildType:Webhook', webhook_buildType - Specifies the type of build this webhook will trigger.

webhook_branchFilter :: Lens' Webhook (Maybe Text)

A regular expression used to determine which repository branches are built when a webhook is triggered. If the name of a branch matches the regular expression, then it is built. If branchFilter is empty, then all branches are built.

It is recommended that you use filterGroups instead of branchFilter.

webhook_lastModifiedSecret :: Lens' Webhook (Maybe UTCTime)

A timestamp that indicates the last time a repository's secret token was modified.

webhook_url :: Lens' Webhook (Maybe Text)

The URL to the webhook.

webhook_secret :: Lens' Webhook (Maybe Text)

The secret token of the associated repository.

A Bitbucket webhook does not support secret.

webhook_filterGroups :: Lens' Webhook (Maybe [[WebhookFilter]])

An array of arrays of WebhookFilter objects used to determine which webhooks are triggered. At least one WebhookFilter in the array must specify EVENT as its type.

For a build to be triggered, at least one filter group in the filterGroups array must pass. For a filter group to pass, each of its filters must pass.

webhook_payloadUrl :: Lens' Webhook (Maybe Text)

The CodeBuild endpoint where webhook events are sent.

webhook_buildType :: Lens' Webhook (Maybe WebhookBuildType)

Specifies the type of build this webhook will trigger.