Network.AWS.CodeBuild.UpdateProjectVisibility

Description

Changes the public visibility for a project. The project's build results, logs, and artifacts are available to the general public. For more information, see Public build projects in the CodeBuild User Guide.

The following should be kept in mind when making your projects public:

All of a project's build results, logs, and artifacts, including builds that were run when the project was private, are available to the general public.

All build logs and artifacts are available to the public. Environment variables, source code, and other sensitive information may have been output to the build logs and artifacts. You must be careful about what information is output to the build logs. Some best practice are:

Do not store sensitive values, especially Amazon Web Services access key IDs and secret access keys, in environment variables. We recommend that you use an Amazon EC2 Systems Manager Parameter Store or Secrets Manager to store sensitive values.

Follow Best practices for using webhooks in the CodeBuild User Guide to limit which entities can trigger a build, and do not store the buildspec in the project itself, to ensure that your webhooks are as secure as possible.

A malicious user can use public builds to distribute malicious artifacts. We recommend that you review all pull requests to verify that the pull request is a legitimate change. We also recommend that you validate any artifacts with their checksums to make sure that the correct artifacts are being downloaded.

Creating a Request

data UpdateProjectVisibility

See: newUpdateProjectVisibility smart constructor.

Constructors

UpdateProjectVisibility'

Fields

resourceAccessRole :: Maybe Text

The ARN of the IAM role that enables CodeBuild to access the CloudWatch Logs and Amazon S3 artifacts for the project's builds.

projectArn :: Text

The Amazon Resource Name (ARN) of the build project.

projectVisibility :: ProjectVisibilityType

newUpdateProjectVisibility

Arguments

:: Text
$sel:projectArn:UpdateProjectVisibility'

-> ProjectVisibilityType
$sel:projectVisibility:UpdateProjectVisibility'

-> UpdateProjectVisibility

Create a value of UpdateProjectVisibility with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceAccessRole:UpdateProjectVisibility', updateProjectVisibility_resourceAccessRole - The ARN of the IAM role that enables CodeBuild to access the CloudWatch Logs and Amazon S3 artifacts for the project's builds.

$sel:projectArn:UpdateProjectVisibility', updateProjectVisibility_projectArn - The Amazon Resource Name (ARN) of the build project.

$sel:projectVisibility:UpdateProjectVisibility', updateProjectVisibility_projectVisibility - Undocumented member.

Request Lenses

updateProjectVisibility_resourceAccessRole :: Lens' UpdateProjectVisibility (Maybe Text)

The ARN of the IAM role that enables CodeBuild to access the CloudWatch Logs and Amazon S3 artifacts for the project's builds.

updateProjectVisibility_projectArn :: Lens' UpdateProjectVisibility Text

The Amazon Resource Name (ARN) of the build project.

updateProjectVisibility_projectVisibility :: Lens' UpdateProjectVisibility ProjectVisibilityType

Undocumented member.

Destructuring the Response

data UpdateProjectVisibilityResponse

See: newUpdateProjectVisibilityResponse smart constructor.

Constructors

UpdateProjectVisibilityResponse'

Fields

publicProjectAlias :: Maybe Text

Contains the project identifier used with the public build APIs.

projectArn :: Maybe Text

The Amazon Resource Name (ARN) of the build project.

projectVisibility :: Maybe ProjectVisibilityType

httpStatus :: Int

The response's http status code.

newUpdateProjectVisibilityResponse

Arguments

:: Int
$sel:httpStatus:UpdateProjectVisibilityResponse'

-> UpdateProjectVisibilityResponse

Create a value of UpdateProjectVisibilityResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:publicProjectAlias:UpdateProjectVisibilityResponse', updateProjectVisibilityResponse_publicProjectAlias - Contains the project identifier used with the public build APIs.

$sel:projectArn:UpdateProjectVisibility', updateProjectVisibilityResponse_projectArn - The Amazon Resource Name (ARN) of the build project.

$sel:projectVisibility:UpdateProjectVisibility', updateProjectVisibilityResponse_projectVisibility - Undocumented member.

$sel:httpStatus:UpdateProjectVisibilityResponse', updateProjectVisibilityResponse_httpStatus - The response's http status code.

Response Lenses

updateProjectVisibilityResponse_publicProjectAlias :: Lens' UpdateProjectVisibilityResponse (Maybe Text)

Contains the project identifier used with the public build APIs.

updateProjectVisibilityResponse_projectArn :: Lens' UpdateProjectVisibilityResponse (Maybe Text)

The Amazon Resource Name (ARN) of the build project.

updateProjectVisibilityResponse_projectVisibility :: Lens' UpdateProjectVisibilityResponse (Maybe ProjectVisibilityType)

Undocumented member.

updateProjectVisibilityResponse_httpStatus :: Lens' UpdateProjectVisibilityResponse Int

The response's http status code.