Network.AWS.CodeBuild.UpdateReportGroup

Description

Updates a report group.

Creating a Request

data UpdateReportGroup

See: newUpdateReportGroup smart constructor.

Constructors

UpdateReportGroup'

Fields

exportConfig :: Maybe ReportExportConfig

Used to specify an updated export type. Valid values are:

S3: The report results are exported to an S3 bucket.

NO_EXPORT: The report results are not exported.

tags :: Maybe [Tag]

An updated list of tag key and value pairs associated with this report group.

These tags are available for use by Amazon Web Services services that support CodeBuild report group tags.

arn :: Text

The ARN of the report group to update.

newUpdateReportGroup

Arguments

:: Text
$sel:arn:UpdateReportGroup'

-> UpdateReportGroup

Create a value of UpdateReportGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:exportConfig:UpdateReportGroup', updateReportGroup_exportConfig - Used to specify an updated export type. Valid values are:

S3: The report results are exported to an S3 bucket.

NO_EXPORT: The report results are not exported.

$sel:tags:UpdateReportGroup', updateReportGroup_tags - An updated list of tag key and value pairs associated with this report group.

These tags are available for use by Amazon Web Services services that support CodeBuild report group tags.

$sel:arn:UpdateReportGroup', updateReportGroup_arn - The ARN of the report group to update.

Request Lenses

updateReportGroup_exportConfig :: Lens' UpdateReportGroup (Maybe ReportExportConfig)

Used to specify an updated export type. Valid values are:

S3: The report results are exported to an S3 bucket.

NO_EXPORT: The report results are not exported.

updateReportGroup_tags :: Lens' UpdateReportGroup (Maybe [Tag])

An updated list of tag key and value pairs associated with this report group.

These tags are available for use by Amazon Web Services services that support CodeBuild report group tags.

updateReportGroup_arn :: Lens' UpdateReportGroup Text

The ARN of the report group to update.

Destructuring the Response

data UpdateReportGroupResponse

See: newUpdateReportGroupResponse smart constructor.

Constructors

UpdateReportGroupResponse'

Fields

reportGroup :: Maybe ReportGroup

Information about the updated report group.

httpStatus :: Int

The response's http status code.

newUpdateReportGroupResponse

Arguments

:: Int
$sel:httpStatus:UpdateReportGroupResponse'

-> UpdateReportGroupResponse

Create a value of UpdateReportGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reportGroup:UpdateReportGroupResponse', updateReportGroupResponse_reportGroup - Information about the updated report group.

$sel:httpStatus:UpdateReportGroupResponse', updateReportGroupResponse_httpStatus - The response's http status code.

Response Lenses

updateReportGroupResponse_reportGroup :: Lens' UpdateReportGroupResponse (Maybe ReportGroup)

Information about the updated report group.

updateReportGroupResponse_httpStatus :: Lens' UpdateReportGroupResponse Int

The response's http status code.