Network.AWS.CodeCommit.BatchGetCommits

Description

Returns information about the contents of one or more commits in a repository.

Creating a Request

data BatchGetCommits

See: newBatchGetCommits smart constructor.

Constructors

BatchGetCommits'

Fields

commitIds :: [Text]

The full commit IDs of the commits to get information about.

You must supply the full SHA IDs of each commit. You cannot use shortened SHA IDs.

repositoryName :: Text

The name of the repository that contains the commits.

newBatchGetCommits

Arguments

:: Text
$sel:repositoryName:BatchGetCommits'

-> BatchGetCommits

Create a value of BatchGetCommits with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:commitIds:BatchGetCommits', batchGetCommits_commitIds - The full commit IDs of the commits to get information about.

You must supply the full SHA IDs of each commit. You cannot use shortened SHA IDs.

$sel:repositoryName:BatchGetCommits', batchGetCommits_repositoryName - The name of the repository that contains the commits.

Request Lenses

batchGetCommits_commitIds :: Lens' BatchGetCommits [Text]

The full commit IDs of the commits to get information about.

You must supply the full SHA IDs of each commit. You cannot use shortened SHA IDs.

batchGetCommits_repositoryName :: Lens' BatchGetCommits Text

The name of the repository that contains the commits.

Destructuring the Response

data BatchGetCommitsResponse

See: newBatchGetCommitsResponse smart constructor.

Constructors

BatchGetCommitsResponse'

Fields

commits :: Maybe [Commit]

An array of commit data type objects, each of which contains information about a specified commit.

errors :: Maybe [BatchGetCommitsError]

Returns any commit IDs for which information could not be found. For example, if one of the commit IDs was a shortened SHA ID or that commit was not found in the specified repository, the ID returns an error object with more information.

httpStatus :: Int

The response's http status code.

newBatchGetCommitsResponse

Arguments

:: Int
$sel:httpStatus:BatchGetCommitsResponse'

-> BatchGetCommitsResponse

Create a value of BatchGetCommitsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:commits:BatchGetCommitsResponse', batchGetCommitsResponse_commits - An array of commit data type objects, each of which contains information about a specified commit.

$sel:errors:BatchGetCommitsResponse', batchGetCommitsResponse_errors - Returns any commit IDs for which information could not be found. For example, if one of the commit IDs was a shortened SHA ID or that commit was not found in the specified repository, the ID returns an error object with more information.

$sel:httpStatus:BatchGetCommitsResponse', batchGetCommitsResponse_httpStatus - The response's http status code.

Response Lenses

batchGetCommitsResponse_commits :: Lens' BatchGetCommitsResponse (Maybe [Commit])

An array of commit data type objects, each of which contains information about a specified commit.

batchGetCommitsResponse_errors :: Lens' BatchGetCommitsResponse (Maybe [BatchGetCommitsError])

Returns any commit IDs for which information could not be found. For example, if one of the commit IDs was a shortened SHA ID or that commit was not found in the specified repository, the ID returns an error object with more information.

batchGetCommitsResponse_httpStatus :: Lens' BatchGetCommitsResponse Int

The response's http status code.