Network.AWS.CodeCommit.CreateCommit

Description

Creates a commit for a repository on the tip of a specified branch.

Creating a Request

data CreateCommit

See: newCreateCommit smart constructor.

Constructors

CreateCommit'

Fields

deleteFiles :: Maybe [DeleteFileEntry]

The files to delete in this commit. These files still exist in earlier commits.

setFileModes :: Maybe [SetFileModeEntry]

The file modes to update for files in this commit.

parentCommitId :: Maybe Text

The ID of the commit that is the parent of the commit you create. Not required if this is an empty repository.

authorName :: Maybe Text

The name of the author who created the commit. This information is used as both the author and committer for the commit.

commitMessage :: Maybe Text

The commit message you want to include in the commit. Commit messages are limited to 256 KB. If no message is specified, a default message is used.

putFiles :: Maybe [PutFileEntry]

The files to add or update in this commit.

email :: Maybe Text

The email address of the person who created the commit.

keepEmptyFolders :: Maybe Bool

If the commit contains deletions, whether to keep a folder or folder structure if the changes leave the folders empty. If true, a ..gitkeep file is created for empty folders. The default is false.

repositoryName :: Text

The name of the repository where you create the commit.

branchName :: Text

The name of the branch where you create the commit.

newCreateCommit

Arguments

:: Text
$sel:repositoryName:CreateCommit'

-> Text
$sel:branchName:CreateCommit'

-> CreateCommit

Create a value of CreateCommit with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deleteFiles:CreateCommit', createCommit_deleteFiles - The files to delete in this commit. These files still exist in earlier commits.

$sel:setFileModes:CreateCommit', createCommit_setFileModes - The file modes to update for files in this commit.

$sel:parentCommitId:CreateCommit', createCommit_parentCommitId - The ID of the commit that is the parent of the commit you create. Not required if this is an empty repository.

$sel:authorName:CreateCommit', createCommit_authorName - The name of the author who created the commit. This information is used as both the author and committer for the commit.

$sel:commitMessage:CreateCommit', createCommit_commitMessage - The commit message you want to include in the commit. Commit messages are limited to 256 KB. If no message is specified, a default message is used.

$sel:putFiles:CreateCommit', createCommit_putFiles - The files to add or update in this commit.

$sel:email:CreateCommit', createCommit_email - The email address of the person who created the commit.

$sel:keepEmptyFolders:CreateCommit', createCommit_keepEmptyFolders - If the commit contains deletions, whether to keep a folder or folder structure if the changes leave the folders empty. If true, a ..gitkeep file is created for empty folders. The default is false.

$sel:repositoryName:CreateCommit', createCommit_repositoryName - The name of the repository where you create the commit.

$sel:branchName:CreateCommit', createCommit_branchName - The name of the branch where you create the commit.

Request Lenses

createCommit_deleteFiles :: Lens' CreateCommit (Maybe [DeleteFileEntry])

The files to delete in this commit. These files still exist in earlier commits.

createCommit_setFileModes :: Lens' CreateCommit (Maybe [SetFileModeEntry])

The file modes to update for files in this commit.

createCommit_parentCommitId :: Lens' CreateCommit (Maybe Text)

The ID of the commit that is the parent of the commit you create. Not required if this is an empty repository.

createCommit_authorName :: Lens' CreateCommit (Maybe Text)

The name of the author who created the commit. This information is used as both the author and committer for the commit.

createCommit_commitMessage :: Lens' CreateCommit (Maybe Text)

The commit message you want to include in the commit. Commit messages are limited to 256 KB. If no message is specified, a default message is used.

createCommit_putFiles :: Lens' CreateCommit (Maybe [PutFileEntry])

The files to add or update in this commit.

createCommit_email :: Lens' CreateCommit (Maybe Text)

The email address of the person who created the commit.

createCommit_keepEmptyFolders :: Lens' CreateCommit (Maybe Bool)

If the commit contains deletions, whether to keep a folder or folder structure if the changes leave the folders empty. If true, a ..gitkeep file is created for empty folders. The default is false.

createCommit_repositoryName :: Lens' CreateCommit Text

The name of the repository where you create the commit.

createCommit_branchName :: Lens' CreateCommit Text

The name of the branch where you create the commit.

Destructuring the Response

data CreateCommitResponse

See: newCreateCommitResponse smart constructor.

Constructors

CreateCommitResponse'

Fields

commitId :: Maybe Text

The full commit ID of the commit that contains your committed file changes.

treeId :: Maybe Text

The full SHA-1 pointer of the tree information for the commit that contains the commited file changes.

filesAdded :: Maybe [FileMetadata]

The files added as part of the committed file changes.

filesDeleted :: Maybe [FileMetadata]

The files deleted as part of the committed file changes.

filesUpdated :: Maybe [FileMetadata]

The files updated as part of the commited file changes.

httpStatus :: Int

The response's http status code.

newCreateCommitResponse

Arguments

:: Int
$sel:httpStatus:CreateCommitResponse'

-> CreateCommitResponse

Create a value of CreateCommitResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:commitId:CreateCommitResponse', createCommitResponse_commitId - The full commit ID of the commit that contains your committed file changes.

$sel:treeId:CreateCommitResponse', createCommitResponse_treeId - The full SHA-1 pointer of the tree information for the commit that contains the commited file changes.

$sel:filesAdded:CreateCommitResponse', createCommitResponse_filesAdded - The files added as part of the committed file changes.

$sel:filesDeleted:CreateCommitResponse', createCommitResponse_filesDeleted - The files deleted as part of the committed file changes.

$sel:filesUpdated:CreateCommitResponse', createCommitResponse_filesUpdated - The files updated as part of the commited file changes.

$sel:httpStatus:CreateCommitResponse', createCommitResponse_httpStatus - The response's http status code.

Response Lenses

createCommitResponse_commitId :: Lens' CreateCommitResponse (Maybe Text)

The full commit ID of the commit that contains your committed file changes.

createCommitResponse_treeId :: Lens' CreateCommitResponse (Maybe Text)

The full SHA-1 pointer of the tree information for the commit that contains the commited file changes.

createCommitResponse_filesAdded :: Lens' CreateCommitResponse (Maybe [FileMetadata])

The files added as part of the committed file changes.

createCommitResponse_filesDeleted :: Lens' CreateCommitResponse (Maybe [FileMetadata])

The files deleted as part of the committed file changes.

createCommitResponse_filesUpdated :: Lens' CreateCommitResponse (Maybe [FileMetadata])

The files updated as part of the commited file changes.

createCommitResponse_httpStatus :: Lens' CreateCommitResponse Int

The response's http status code.