Network.AWS.CodeCommit.CreatePullRequest

Description

Creates a pull request in the specified repository.

Creating a Request

data CreatePullRequest

See: newCreatePullRequest smart constructor.

Constructors

CreatePullRequest'

Fields

description :: Maybe Text

A description of the pull request.

clientRequestToken :: Maybe Text

A unique, client-generated idempotency token that, when provided in a request, ensures the request cannot be repeated with a changed parameter. If a request is received with the same parameters and a token is included, the request returns information about the initial request that used that token.

The AWS SDKs prepopulate client request tokens. If you are using an AWS SDK, an idempotency token is created for you.

title :: Text

The title of the pull request. This title is used to identify the pull request to other users in the repository.

targets :: [Target]

The targets for the pull request, including the source of the code to be reviewed (the source branch) and the destination where the creator of the pull request intends the code to be merged after the pull request is closed (the destination branch).

newCreatePullRequest

Arguments

:: Text
$sel:title:CreatePullRequest'

-> CreatePullRequest

Create a value of CreatePullRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:CreatePullRequest', createPullRequest_description - A description of the pull request.

$sel:clientRequestToken:CreatePullRequest', createPullRequest_clientRequestToken - A unique, client-generated idempotency token that, when provided in a request, ensures the request cannot be repeated with a changed parameter. If a request is received with the same parameters and a token is included, the request returns information about the initial request that used that token.

The AWS SDKs prepopulate client request tokens. If you are using an AWS SDK, an idempotency token is created for you.

$sel:title:CreatePullRequest', createPullRequest_title - The title of the pull request. This title is used to identify the pull request to other users in the repository.

$sel:targets:CreatePullRequest', createPullRequest_targets - The targets for the pull request, including the source of the code to be reviewed (the source branch) and the destination where the creator of the pull request intends the code to be merged after the pull request is closed (the destination branch).

Request Lenses

createPullRequest_description :: Lens' CreatePullRequest (Maybe Text)

A description of the pull request.

createPullRequest_clientRequestToken :: Lens' CreatePullRequest (Maybe Text)

A unique, client-generated idempotency token that, when provided in a request, ensures the request cannot be repeated with a changed parameter. If a request is received with the same parameters and a token is included, the request returns information about the initial request that used that token.

The AWS SDKs prepopulate client request tokens. If you are using an AWS SDK, an idempotency token is created for you.

createPullRequest_title :: Lens' CreatePullRequest Text

The title of the pull request. This title is used to identify the pull request to other users in the repository.

createPullRequest_targets :: Lens' CreatePullRequest [Target]

The targets for the pull request, including the source of the code to be reviewed (the source branch) and the destination where the creator of the pull request intends the code to be merged after the pull request is closed (the destination branch).

Destructuring the Response

data CreatePullRequestResponse

See: newCreatePullRequestResponse smart constructor.

Constructors

CreatePullRequestResponse'

Fields

httpStatus :: Int

The response's http status code.

pullRequest :: PullRequest

Information about the newly created pull request.

newCreatePullRequestResponse

Arguments

:: Int
$sel:httpStatus:CreatePullRequestResponse'

-> PullRequest
$sel:pullRequest:CreatePullRequestResponse'

-> CreatePullRequestResponse

Create a value of CreatePullRequestResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreatePullRequestResponse', createPullRequestResponse_httpStatus - The response's http status code.

$sel:pullRequest:CreatePullRequestResponse', createPullRequestResponse_pullRequest - Information about the newly created pull request.

Response Lenses

createPullRequestResponse_httpStatus :: Lens' CreatePullRequestResponse Int

The response's http status code.

createPullRequestResponse_pullRequest :: Lens' CreatePullRequestResponse PullRequest

Information about the newly created pull request.