Network.AWS.CodeCommit.CreatePullRequestApprovalRule

Description

Creates an approval rule for a pull request.

Creating a Request

data CreatePullRequestApprovalRule

See: newCreatePullRequestApprovalRule smart constructor.

Constructors

CreatePullRequestApprovalRule'

Fields

pullRequestId :: Text

The system-generated ID of the pull request for which you want to create the approval rule.

approvalRuleName :: Text

The name for the approval rule.

approvalRuleContent :: Text

The content of the approval rule, including the number of approvals needed and the structure of an approval pool defined for approvals, if any. For more information about approval pools, see the AWS CodeCommit User Guide.

When you create the content of the approval rule, you can specify approvers in an approval pool in one of two ways:

CodeCommitApprovers: This option only requires an AWS account and a resource. It can be used for both IAM users and federated access users whose name matches the provided resource name. This is a very powerful option that offers a great deal of flexibility. For example, if you specify the AWS account 123456789012 and Mary_Major, all of the following would be counted as approvals coming from that user:

An IAM user in the account (arn:aws:iam::123456789012:user/Mary_Major)

A federated user identified in IAM as Mary_Major (arn:aws:sts::123456789012:federated-user/Mary_Major)

This option does not recognize an active session of someone assuming the role of CodeCommitReview with a role session name of Mary_Major (arn:aws:sts::123456789012:assumed-role/CodeCommitReview/Mary_Major) unless you include a wildcard (*Mary_Major).

Fully qualified ARN: This option allows you to specify the fully qualified Amazon Resource Name (ARN) of the IAM user or role.

For more information about IAM ARNs, wildcards, and formats, see IAM Identifiers in the IAM User Guide.

newCreatePullRequestApprovalRule

Arguments

:: Text
$sel:pullRequestId:CreatePullRequestApprovalRule'

-> Text
$sel:approvalRuleName:CreatePullRequestApprovalRule'

-> Text
$sel:approvalRuleContent:CreatePullRequestApprovalRule'

-> CreatePullRequestApprovalRule

Create a value of CreatePullRequestApprovalRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pullRequestId:CreatePullRequestApprovalRule', createPullRequestApprovalRule_pullRequestId - The system-generated ID of the pull request for which you want to create the approval rule.

$sel:approvalRuleName:CreatePullRequestApprovalRule', createPullRequestApprovalRule_approvalRuleName - The name for the approval rule.

$sel:approvalRuleContent:CreatePullRequestApprovalRule', createPullRequestApprovalRule_approvalRuleContent - The content of the approval rule, including the number of approvals needed and the structure of an approval pool defined for approvals, if any. For more information about approval pools, see the AWS CodeCommit User Guide.

When you create the content of the approval rule, you can specify approvers in an approval pool in one of two ways:

CodeCommitApprovers: This option only requires an AWS account and a resource. It can be used for both IAM users and federated access users whose name matches the provided resource name. This is a very powerful option that offers a great deal of flexibility. For example, if you specify the AWS account 123456789012 and Mary_Major, all of the following would be counted as approvals coming from that user:

An IAM user in the account (arn:aws:iam::123456789012:user/Mary_Major)

A federated user identified in IAM as Mary_Major (arn:aws:sts::123456789012:federated-user/Mary_Major)

This option does not recognize an active session of someone assuming the role of CodeCommitReview with a role session name of Mary_Major (arn:aws:sts::123456789012:assumed-role/CodeCommitReview/Mary_Major) unless you include a wildcard (*Mary_Major).

Fully qualified ARN: This option allows you to specify the fully qualified Amazon Resource Name (ARN) of the IAM user or role.

For more information about IAM ARNs, wildcards, and formats, see IAM Identifiers in the IAM User Guide.

Request Lenses

createPullRequestApprovalRule_pullRequestId :: Lens' CreatePullRequestApprovalRule Text

The system-generated ID of the pull request for which you want to create the approval rule.

createPullRequestApprovalRule_approvalRuleName :: Lens' CreatePullRequestApprovalRule Text

The name for the approval rule.

createPullRequestApprovalRule_approvalRuleContent :: Lens' CreatePullRequestApprovalRule Text

The content of the approval rule, including the number of approvals needed and the structure of an approval pool defined for approvals, if any. For more information about approval pools, see the AWS CodeCommit User Guide.

When you create the content of the approval rule, you can specify approvers in an approval pool in one of two ways:

CodeCommitApprovers: This option only requires an AWS account and a resource. It can be used for both IAM users and federated access users whose name matches the provided resource name. This is a very powerful option that offers a great deal of flexibility. For example, if you specify the AWS account 123456789012 and Mary_Major, all of the following would be counted as approvals coming from that user:

An IAM user in the account (arn:aws:iam::123456789012:user/Mary_Major)

A federated user identified in IAM as Mary_Major (arn:aws:sts::123456789012:federated-user/Mary_Major)

This option does not recognize an active session of someone assuming the role of CodeCommitReview with a role session name of Mary_Major (arn:aws:sts::123456789012:assumed-role/CodeCommitReview/Mary_Major) unless you include a wildcard (*Mary_Major).

Fully qualified ARN: This option allows you to specify the fully qualified Amazon Resource Name (ARN) of the IAM user or role.

For more information about IAM ARNs, wildcards, and formats, see IAM Identifiers in the IAM User Guide.

Destructuring the Response

data CreatePullRequestApprovalRuleResponse

See: newCreatePullRequestApprovalRuleResponse smart constructor.

Constructors

CreatePullRequestApprovalRuleResponse'

Fields

httpStatus :: Int

The response's http status code.

approvalRule :: ApprovalRule

Information about the created approval rule.

newCreatePullRequestApprovalRuleResponse

Arguments

:: Int
$sel:httpStatus:CreatePullRequestApprovalRuleResponse'

-> ApprovalRule
$sel:approvalRule:CreatePullRequestApprovalRuleResponse'

-> CreatePullRequestApprovalRuleResponse

Create a value of CreatePullRequestApprovalRuleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreatePullRequestApprovalRuleResponse', createPullRequestApprovalRuleResponse_httpStatus - The response's http status code.

$sel:approvalRule:CreatePullRequestApprovalRuleResponse', createPullRequestApprovalRuleResponse_approvalRule - Information about the created approval rule.

Response Lenses

createPullRequestApprovalRuleResponse_httpStatus :: Lens' CreatePullRequestApprovalRuleResponse Int

The response's http status code.

createPullRequestApprovalRuleResponse_approvalRule :: Lens' CreatePullRequestApprovalRuleResponse ApprovalRule

Information about the created approval rule.