Network.AWS.CodeCommit.DeleteBranch

Description

Deletes a branch from a repository, unless that branch is the default branch for the repository.

Creating a Request

data DeleteBranch

Represents the input of a delete branch operation.

See: newDeleteBranch smart constructor.

Constructors

DeleteBranch'

Fields

repositoryName :: Text

The name of the repository that contains the branch to be deleted.

branchName :: Text

The name of the branch to delete.

newDeleteBranch

Arguments

:: Text
$sel:repositoryName:DeleteBranch'

-> Text
$sel:branchName:DeleteBranch'

-> DeleteBranch

Create a value of DeleteBranch with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:repositoryName:DeleteBranch', deleteBranch_repositoryName - The name of the repository that contains the branch to be deleted.

$sel:branchName:DeleteBranch', deleteBranch_branchName - The name of the branch to delete.

Request Lenses

deleteBranch_repositoryName :: Lens' DeleteBranch Text

The name of the repository that contains the branch to be deleted.

deleteBranch_branchName :: Lens' DeleteBranch Text

The name of the branch to delete.

Destructuring the Response

data DeleteBranchResponse

Represents the output of a delete branch operation.

See: newDeleteBranchResponse smart constructor.

Constructors

DeleteBranchResponse'

Fields

deletedBranch :: Maybe BranchInfo

Information about the branch deleted by the operation, including the branch name and the commit ID that was the tip of the branch.

httpStatus :: Int

The response's http status code.

newDeleteBranchResponse

Arguments

:: Int
$sel:httpStatus:DeleteBranchResponse'

-> DeleteBranchResponse

Create a value of DeleteBranchResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deletedBranch:DeleteBranchResponse', deleteBranchResponse_deletedBranch - Information about the branch deleted by the operation, including the branch name and the commit ID that was the tip of the branch.

$sel:httpStatus:DeleteBranchResponse', deleteBranchResponse_httpStatus - The response's http status code.

Response Lenses

deleteBranchResponse_deletedBranch :: Lens' DeleteBranchResponse (Maybe BranchInfo)

Information about the branch deleted by the operation, including the branch name and the commit ID that was the tip of the branch.

deleteBranchResponse_httpStatus :: Lens' DeleteBranchResponse Int

The response's http status code.