Network.AWS.CodeCommit.DeleteFile

Description

Deletes a specified file from a specified branch. A commit is created on the branch that contains the revision. The file still exists in the commits earlier to the commit that contains the deletion.

Creating a Request

data DeleteFile

See: newDeleteFile smart constructor.

Constructors

DeleteFile'

Fields

commitMessage :: Maybe Text

The commit message you want to include as part of deleting the file. Commit messages are limited to 256 KB. If no message is specified, a default message is used.

name :: Maybe Text

The name of the author of the commit that deletes the file. If no name is specified, the user's ARN is used as the author name and committer name.

email :: Maybe Text

The email address for the commit that deletes the file. If no email address is specified, the email address is left blank.

keepEmptyFolders :: Maybe Bool

If a file is the only object in the folder or directory, specifies whether to delete the folder or directory that contains the file. By default, empty folders are deleted. This includes empty folders that are part of the directory structure. For example, if the path to a file is dir1/dir2/dir3/dir4, and dir2 and dir3 are empty, deleting the last file in dir4 also deletes the empty folders dir4, dir3, and dir2.

repositoryName :: Text

The name of the repository that contains the file to delete.

branchName :: Text

The name of the branch where the commit that deletes the file is made.

filePath :: Text

The fully qualified path to the file that to be deleted, including the full name and extension of that file. For example, /examples/file.md is a fully qualified path to a file named file.md in a folder named examples.

parentCommitId :: Text

The ID of the commit that is the tip of the branch where you want to create the commit that deletes the file. This must be the HEAD commit for the branch. The commit that deletes the file is created from this commit ID.

newDeleteFile

Arguments

:: Text
$sel:repositoryName:DeleteFile'

-> Text
$sel:branchName:DeleteFile'

-> Text
$sel:filePath:DeleteFile'

-> Text
$sel:parentCommitId:DeleteFile'

-> DeleteFile

Create a value of DeleteFile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:commitMessage:DeleteFile', deleteFile_commitMessage - The commit message you want to include as part of deleting the file. Commit messages are limited to 256 KB. If no message is specified, a default message is used.

$sel:name:DeleteFile', deleteFile_name - The name of the author of the commit that deletes the file. If no name is specified, the user's ARN is used as the author name and committer name.

$sel:email:DeleteFile', deleteFile_email - The email address for the commit that deletes the file. If no email address is specified, the email address is left blank.

$sel:keepEmptyFolders:DeleteFile', deleteFile_keepEmptyFolders - If a file is the only object in the folder or directory, specifies whether to delete the folder or directory that contains the file. By default, empty folders are deleted. This includes empty folders that are part of the directory structure. For example, if the path to a file is dir1/dir2/dir3/dir4, and dir2 and dir3 are empty, deleting the last file in dir4 also deletes the empty folders dir4, dir3, and dir2.

$sel:repositoryName:DeleteFile', deleteFile_repositoryName - The name of the repository that contains the file to delete.

$sel:branchName:DeleteFile', deleteFile_branchName - The name of the branch where the commit that deletes the file is made.

$sel:filePath:DeleteFile', deleteFile_filePath - The fully qualified path to the file that to be deleted, including the full name and extension of that file. For example, /examples/file.md is a fully qualified path to a file named file.md in a folder named examples.

$sel:parentCommitId:DeleteFile', deleteFile_parentCommitId - The ID of the commit that is the tip of the branch where you want to create the commit that deletes the file. This must be the HEAD commit for the branch. The commit that deletes the file is created from this commit ID.

Request Lenses

deleteFile_commitMessage :: Lens' DeleteFile (Maybe Text)

The commit message you want to include as part of deleting the file. Commit messages are limited to 256 KB. If no message is specified, a default message is used.

deleteFile_name :: Lens' DeleteFile (Maybe Text)

The name of the author of the commit that deletes the file. If no name is specified, the user's ARN is used as the author name and committer name.

deleteFile_email :: Lens' DeleteFile (Maybe Text)

The email address for the commit that deletes the file. If no email address is specified, the email address is left blank.

deleteFile_keepEmptyFolders :: Lens' DeleteFile (Maybe Bool)

If a file is the only object in the folder or directory, specifies whether to delete the folder or directory that contains the file. By default, empty folders are deleted. This includes empty folders that are part of the directory structure. For example, if the path to a file is dir1/dir2/dir3/dir4, and dir2 and dir3 are empty, deleting the last file in dir4 also deletes the empty folders dir4, dir3, and dir2.

deleteFile_repositoryName :: Lens' DeleteFile Text

The name of the repository that contains the file to delete.

deleteFile_branchName :: Lens' DeleteFile Text

The name of the branch where the commit that deletes the file is made.

deleteFile_filePath :: Lens' DeleteFile Text

The fully qualified path to the file that to be deleted, including the full name and extension of that file. For example, /examples/file.md is a fully qualified path to a file named file.md in a folder named examples.

deleteFile_parentCommitId :: Lens' DeleteFile Text

The ID of the commit that is the tip of the branch where you want to create the commit that deletes the file. This must be the HEAD commit for the branch. The commit that deletes the file is created from this commit ID.

Destructuring the Response

data DeleteFileResponse

See: newDeleteFileResponse smart constructor.

Constructors

DeleteFileResponse'

Fields

httpStatus :: Int

The response's http status code.

commitId :: Text

The full commit ID of the commit that contains the change that deletes the file.

blobId :: Text

The blob ID removed from the tree as part of deleting the file.

treeId :: Text

The full SHA-1 pointer of the tree information for the commit that contains the delete file change.

filePath :: Text

The fully qualified path to the file to be deleted, including the full name and extension of that file.

newDeleteFileResponse

Arguments

:: Int
$sel:httpStatus:DeleteFileResponse'

-> Text
$sel:commitId:DeleteFileResponse'

-> Text
$sel:blobId:DeleteFileResponse'

-> Text
$sel:treeId:DeleteFileResponse'

-> Text
$sel:filePath:DeleteFile'

-> DeleteFileResponse

Create a value of DeleteFileResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteFileResponse', deleteFileResponse_httpStatus - The response's http status code.

$sel:commitId:DeleteFileResponse', deleteFileResponse_commitId - The full commit ID of the commit that contains the change that deletes the file.

$sel:blobId:DeleteFileResponse', deleteFileResponse_blobId - The blob ID removed from the tree as part of deleting the file.

$sel:treeId:DeleteFileResponse', deleteFileResponse_treeId - The full SHA-1 pointer of the tree information for the commit that contains the delete file change.

$sel:filePath:DeleteFile', deleteFileResponse_filePath - The fully qualified path to the file to be deleted, including the full name and extension of that file.

Response Lenses

deleteFileResponse_httpStatus :: Lens' DeleteFileResponse Int

The response's http status code.

deleteFileResponse_commitId :: Lens' DeleteFileResponse Text

The full commit ID of the commit that contains the change that deletes the file.

deleteFileResponse_blobId :: Lens' DeleteFileResponse Text

The blob ID removed from the tree as part of deleting the file.

deleteFileResponse_treeId :: Lens' DeleteFileResponse Text

The full SHA-1 pointer of the tree information for the commit that contains the delete file change.

deleteFileResponse_filePath :: Lens' DeleteFileResponse Text

The fully qualified path to the file to be deleted, including the full name and extension of that file.