Network.AWS.CodeCommit.DeleteRepository

Description

Deletes a repository. If a specified repository was already deleted, a null repository ID is returned.

Deleting a repository also deletes all associated objects and metadata. After a repository is deleted, all future push calls to the deleted repository fail.

Creating a Request

data DeleteRepository

Represents the input of a delete repository operation.

See: newDeleteRepository smart constructor.

Constructors

DeleteRepository'

Fields

repositoryName :: Text

The name of the repository to delete.

newDeleteRepository

Arguments

:: Text
$sel:repositoryName:DeleteRepository'

-> DeleteRepository

Create a value of DeleteRepository with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:repositoryName:DeleteRepository', deleteRepository_repositoryName - The name of the repository to delete.

Request Lenses

deleteRepository_repositoryName :: Lens' DeleteRepository Text

The name of the repository to delete.

Destructuring the Response

data DeleteRepositoryResponse

Represents the output of a delete repository operation.

See: newDeleteRepositoryResponse smart constructor.

Constructors

DeleteRepositoryResponse'

Fields

repositoryId :: Maybe Text

The ID of the repository that was deleted.

httpStatus :: Int

The response's http status code.

newDeleteRepositoryResponse

Arguments

:: Int
$sel:httpStatus:DeleteRepositoryResponse'

-> DeleteRepositoryResponse

Create a value of DeleteRepositoryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:repositoryId:DeleteRepositoryResponse', deleteRepositoryResponse_repositoryId - The ID of the repository that was deleted.

$sel:httpStatus:DeleteRepositoryResponse', deleteRepositoryResponse_httpStatus - The response's http status code.

Response Lenses

deleteRepositoryResponse_repositoryId :: Lens' DeleteRepositoryResponse (Maybe Text)

The ID of the repository that was deleted.

deleteRepositoryResponse_httpStatus :: Lens' DeleteRepositoryResponse Int

The response's http status code.