Network.AWS.CodeCommit.DescribeMergeConflicts

Description

Returns information about one or more merge conflicts in the attempted merge of two commit specifiers using the squash or three-way merge strategy. If the merge option for the attempted merge is specified as FAST_FORWARD_MERGE, an exception is thrown.

Creating a Request

data DescribeMergeConflicts

See: newDescribeMergeConflicts smart constructor.

Constructors

DescribeMergeConflicts'

Fields

maxMergeHunks :: Maybe Int

The maximum number of merge hunks to include in the output.

nextToken :: Maybe Text

An enumeration token that, when provided in a request, returns the next batch of the results.

conflictDetailLevel :: Maybe ConflictDetailLevelTypeEnum

The level of conflict detail to use. If unspecified, the default FILE_LEVEL is used, which returns a not-mergeable result if the same file has differences in both branches. If LINE_LEVEL is specified, a conflict is considered not mergeable if the same file in both branches has differences on the same line.

conflictResolutionStrategy :: Maybe ConflictResolutionStrategyTypeEnum

Specifies which branch to use when resolving conflicts, or whether to attempt automatically merging two versions of a file. The default is NONE, which requires any conflicts to be resolved manually before the merge operation is successful.

repositoryName :: Text

The name of the repository where you want to get information about a merge conflict.

destinationCommitSpecifier :: Text

The branch, tag, HEAD, or other fully qualified reference used to identify a commit (for example, a branch name or a full commit ID).

sourceCommitSpecifier :: Text

The branch, tag, HEAD, or other fully qualified reference used to identify a commit (for example, a branch name or a full commit ID).

mergeOption :: MergeOptionTypeEnum

The merge option or strategy you want to use to merge the code.

filePath :: Text

The path of the target files used to describe the conflicts.

newDescribeMergeConflicts

Arguments

:: Text
$sel:repositoryName:DescribeMergeConflicts'

-> Text
$sel:destinationCommitSpecifier:DescribeMergeConflicts'

-> Text
$sel:sourceCommitSpecifier:DescribeMergeConflicts'

-> MergeOptionTypeEnum
$sel:mergeOption:DescribeMergeConflicts'

-> Text
$sel:filePath:DescribeMergeConflicts'

-> DescribeMergeConflicts

Create a value of DescribeMergeConflicts with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxMergeHunks:DescribeMergeConflicts', describeMergeConflicts_maxMergeHunks - The maximum number of merge hunks to include in the output.

$sel:nextToken:DescribeMergeConflicts', describeMergeConflicts_nextToken - An enumeration token that, when provided in a request, returns the next batch of the results.

$sel:conflictDetailLevel:DescribeMergeConflicts', describeMergeConflicts_conflictDetailLevel - The level of conflict detail to use. If unspecified, the default FILE_LEVEL is used, which returns a not-mergeable result if the same file has differences in both branches. If LINE_LEVEL is specified, a conflict is considered not mergeable if the same file in both branches has differences on the same line.

$sel:conflictResolutionStrategy:DescribeMergeConflicts', describeMergeConflicts_conflictResolutionStrategy - Specifies which branch to use when resolving conflicts, or whether to attempt automatically merging two versions of a file. The default is NONE, which requires any conflicts to be resolved manually before the merge operation is successful.

$sel:repositoryName:DescribeMergeConflicts', describeMergeConflicts_repositoryName - The name of the repository where you want to get information about a merge conflict.

$sel:destinationCommitSpecifier:DescribeMergeConflicts', describeMergeConflicts_destinationCommitSpecifier - The branch, tag, HEAD, or other fully qualified reference used to identify a commit (for example, a branch name or a full commit ID).

$sel:sourceCommitSpecifier:DescribeMergeConflicts', describeMergeConflicts_sourceCommitSpecifier - The branch, tag, HEAD, or other fully qualified reference used to identify a commit (for example, a branch name or a full commit ID).

$sel:mergeOption:DescribeMergeConflicts', describeMergeConflicts_mergeOption - The merge option or strategy you want to use to merge the code.

$sel:filePath:DescribeMergeConflicts', describeMergeConflicts_filePath - The path of the target files used to describe the conflicts.

Request Lenses

describeMergeConflicts_maxMergeHunks :: Lens' DescribeMergeConflicts (Maybe Int)

The maximum number of merge hunks to include in the output.

describeMergeConflicts_nextToken :: Lens' DescribeMergeConflicts (Maybe Text)

An enumeration token that, when provided in a request, returns the next batch of the results.

describeMergeConflicts_conflictDetailLevel :: Lens' DescribeMergeConflicts (Maybe ConflictDetailLevelTypeEnum)

The level of conflict detail to use. If unspecified, the default FILE_LEVEL is used, which returns a not-mergeable result if the same file has differences in both branches. If LINE_LEVEL is specified, a conflict is considered not mergeable if the same file in both branches has differences on the same line.

describeMergeConflicts_conflictResolutionStrategy :: Lens' DescribeMergeConflicts (Maybe ConflictResolutionStrategyTypeEnum)

Specifies which branch to use when resolving conflicts, or whether to attempt automatically merging two versions of a file. The default is NONE, which requires any conflicts to be resolved manually before the merge operation is successful.

describeMergeConflicts_repositoryName :: Lens' DescribeMergeConflicts Text

The name of the repository where you want to get information about a merge conflict.

describeMergeConflicts_destinationCommitSpecifier :: Lens' DescribeMergeConflicts Text

The branch, tag, HEAD, or other fully qualified reference used to identify a commit (for example, a branch name or a full commit ID).

describeMergeConflicts_sourceCommitSpecifier :: Lens' DescribeMergeConflicts Text

The branch, tag, HEAD, or other fully qualified reference used to identify a commit (for example, a branch name or a full commit ID).

describeMergeConflicts_mergeOption :: Lens' DescribeMergeConflicts MergeOptionTypeEnum

The merge option or strategy you want to use to merge the code.

describeMergeConflicts_filePath :: Lens' DescribeMergeConflicts Text

The path of the target files used to describe the conflicts.

Destructuring the Response

data DescribeMergeConflictsResponse

See: newDescribeMergeConflictsResponse smart constructor.

Constructors

DescribeMergeConflictsResponse'

Fields

nextToken :: Maybe Text

An enumeration token that can be used in a request to return the next batch of the results.

baseCommitId :: Maybe Text

The commit ID of the merge base.

httpStatus :: Int

The response's http status code.

conflictMetadata :: ConflictMetadata

Contains metadata about the conflicts found in the merge.

mergeHunks :: [MergeHunk]

A list of merge hunks of the differences between the files or lines.

destinationCommitId :: Text

The commit ID of the destination commit specifier that was used in the merge evaluation.

sourceCommitId :: Text

The commit ID of the source commit specifier that was used in the merge evaluation.

newDescribeMergeConflictsResponse

Arguments

:: Int
$sel:httpStatus:DescribeMergeConflictsResponse'

-> ConflictMetadata
$sel:conflictMetadata:DescribeMergeConflictsResponse'

-> Text
$sel:destinationCommitId:DescribeMergeConflictsResponse'

-> Text
$sel:sourceCommitId:DescribeMergeConflictsResponse'

-> DescribeMergeConflictsResponse

Create a value of DescribeMergeConflictsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeMergeConflicts', describeMergeConflictsResponse_nextToken - An enumeration token that can be used in a request to return the next batch of the results.

$sel:baseCommitId:DescribeMergeConflictsResponse', describeMergeConflictsResponse_baseCommitId - The commit ID of the merge base.

$sel:httpStatus:DescribeMergeConflictsResponse', describeMergeConflictsResponse_httpStatus - The response's http status code.

$sel:conflictMetadata:DescribeMergeConflictsResponse', describeMergeConflictsResponse_conflictMetadata - Contains metadata about the conflicts found in the merge.

$sel:mergeHunks:DescribeMergeConflictsResponse', describeMergeConflictsResponse_mergeHunks - A list of merge hunks of the differences between the files or lines.

$sel:destinationCommitId:DescribeMergeConflictsResponse', describeMergeConflictsResponse_destinationCommitId - The commit ID of the destination commit specifier that was used in the merge evaluation.

$sel:sourceCommitId:DescribeMergeConflictsResponse', describeMergeConflictsResponse_sourceCommitId - The commit ID of the source commit specifier that was used in the merge evaluation.

Response Lenses

describeMergeConflictsResponse_nextToken :: Lens' DescribeMergeConflictsResponse (Maybe Text)

An enumeration token that can be used in a request to return the next batch of the results.

describeMergeConflictsResponse_baseCommitId :: Lens' DescribeMergeConflictsResponse (Maybe Text)

The commit ID of the merge base.

describeMergeConflictsResponse_httpStatus :: Lens' DescribeMergeConflictsResponse Int

The response's http status code.

describeMergeConflictsResponse_conflictMetadata :: Lens' DescribeMergeConflictsResponse ConflictMetadata

Contains metadata about the conflicts found in the merge.

describeMergeConflictsResponse_mergeHunks :: Lens' DescribeMergeConflictsResponse [MergeHunk]

A list of merge hunks of the differences between the files or lines.

describeMergeConflictsResponse_destinationCommitId :: Lens' DescribeMergeConflictsResponse Text

The commit ID of the destination commit specifier that was used in the merge evaluation.

describeMergeConflictsResponse_sourceCommitId :: Lens' DescribeMergeConflictsResponse Text

The commit ID of the source commit specifier that was used in the merge evaluation.