Network.AWS.CodeCommit.EvaluatePullRequestApprovalRules

Description

Evaluates whether a pull request has met all the conditions specified in its associated approval rules.

Creating a Request

data EvaluatePullRequestApprovalRules

See: newEvaluatePullRequestApprovalRules smart constructor.

Constructors

EvaluatePullRequestApprovalRules'

Fields

pullRequestId :: Text

The system-generated ID of the pull request you want to evaluate.

revisionId :: Text

The system-generated ID for the pull request revision. To retrieve the most recent revision ID for a pull request, use GetPullRequest.

newEvaluatePullRequestApprovalRules

Arguments

:: Text
$sel:pullRequestId:EvaluatePullRequestApprovalRules'

-> Text
$sel:revisionId:EvaluatePullRequestApprovalRules'

-> EvaluatePullRequestApprovalRules

Create a value of EvaluatePullRequestApprovalRules with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pullRequestId:EvaluatePullRequestApprovalRules', evaluatePullRequestApprovalRules_pullRequestId - The system-generated ID of the pull request you want to evaluate.

$sel:revisionId:EvaluatePullRequestApprovalRules', evaluatePullRequestApprovalRules_revisionId - The system-generated ID for the pull request revision. To retrieve the most recent revision ID for a pull request, use GetPullRequest.

Request Lenses

evaluatePullRequestApprovalRules_pullRequestId :: Lens' EvaluatePullRequestApprovalRules Text

The system-generated ID of the pull request you want to evaluate.

evaluatePullRequestApprovalRules_revisionId :: Lens' EvaluatePullRequestApprovalRules Text

The system-generated ID for the pull request revision. To retrieve the most recent revision ID for a pull request, use GetPullRequest.

Destructuring the Response

data EvaluatePullRequestApprovalRulesResponse

See: newEvaluatePullRequestApprovalRulesResponse smart constructor.

Constructors

EvaluatePullRequestApprovalRulesResponse'

Fields

httpStatus :: Int

The response's http status code.

evaluation :: Evaluation

The result of the evaluation, including the names of the rules whose conditions have been met (if any), the names of the rules whose conditions have not been met (if any), whether the pull request is in the approved state, and whether the pull request approval rule has been set aside by an override.

newEvaluatePullRequestApprovalRulesResponse

Arguments

:: Int
$sel:httpStatus:EvaluatePullRequestApprovalRulesResponse'

-> Evaluation
$sel:evaluation:EvaluatePullRequestApprovalRulesResponse'

-> EvaluatePullRequestApprovalRulesResponse

Create a value of EvaluatePullRequestApprovalRulesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:EvaluatePullRequestApprovalRulesResponse', evaluatePullRequestApprovalRulesResponse_httpStatus - The response's http status code.

$sel:evaluation:EvaluatePullRequestApprovalRulesResponse', evaluatePullRequestApprovalRulesResponse_evaluation - The result of the evaluation, including the names of the rules whose conditions have been met (if any), the names of the rules whose conditions have not been met (if any), whether the pull request is in the approved state, and whether the pull request approval rule has been set aside by an override.

Response Lenses

evaluatePullRequestApprovalRulesResponse_httpStatus :: Lens' EvaluatePullRequestApprovalRulesResponse Int

The response's http status code.

evaluatePullRequestApprovalRulesResponse_evaluation :: Lens' EvaluatePullRequestApprovalRulesResponse Evaluation

The result of the evaluation, including the names of the rules whose conditions have been met (if any), the names of the rules whose conditions have not been met (if any), whether the pull request is in the approved state, and whether the pull request approval rule has been set aside by an override.