Network.AWS.CodeCommit.GetCommentsForComparedCommit

Description

Returns information about comments made on the comparison between two commits.

Reaction counts might include numbers from user identities who were deleted after the reaction was made. For a count of reactions from active identities, use GetCommentReactions.

This operation returns paginated results.

Creating a Request

data GetCommentsForComparedCommit

See: newGetCommentsForComparedCommit smart constructor.

Constructors

GetCommentsForComparedCommit'

Fields

nextToken :: Maybe Text

An enumeration token that when provided in a request, returns the next batch of the results.

beforeCommitId :: Maybe Text

To establish the directionality of the comparison, the full commit ID of the before commit.

maxResults :: Maybe Int

A non-zero, non-negative integer used to limit the number of returned results. The default is 100 comments, but you can configure up to 500.

repositoryName :: Text

The name of the repository where you want to compare commits.

afterCommitId :: Text

To establish the directionality of the comparison, the full commit ID of the after commit.

newGetCommentsForComparedCommit

Arguments

:: Text
$sel:repositoryName:GetCommentsForComparedCommit'

-> Text
$sel:afterCommitId:GetCommentsForComparedCommit'

-> GetCommentsForComparedCommit

Create a value of GetCommentsForComparedCommit with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetCommentsForComparedCommit', getCommentsForComparedCommit_nextToken - An enumeration token that when provided in a request, returns the next batch of the results.

$sel:beforeCommitId:GetCommentsForComparedCommit', getCommentsForComparedCommit_beforeCommitId - To establish the directionality of the comparison, the full commit ID of the before commit.

$sel:maxResults:GetCommentsForComparedCommit', getCommentsForComparedCommit_maxResults - A non-zero, non-negative integer used to limit the number of returned results. The default is 100 comments, but you can configure up to 500.

$sel:repositoryName:GetCommentsForComparedCommit', getCommentsForComparedCommit_repositoryName - The name of the repository where you want to compare commits.

$sel:afterCommitId:GetCommentsForComparedCommit', getCommentsForComparedCommit_afterCommitId - To establish the directionality of the comparison, the full commit ID of the after commit.

Request Lenses

getCommentsForComparedCommit_nextToken :: Lens' GetCommentsForComparedCommit (Maybe Text)

An enumeration token that when provided in a request, returns the next batch of the results.

getCommentsForComparedCommit_beforeCommitId :: Lens' GetCommentsForComparedCommit (Maybe Text)

To establish the directionality of the comparison, the full commit ID of the before commit.

getCommentsForComparedCommit_maxResults :: Lens' GetCommentsForComparedCommit (Maybe Int)

A non-zero, non-negative integer used to limit the number of returned results. The default is 100 comments, but you can configure up to 500.

getCommentsForComparedCommit_repositoryName :: Lens' GetCommentsForComparedCommit Text

The name of the repository where you want to compare commits.

getCommentsForComparedCommit_afterCommitId :: Lens' GetCommentsForComparedCommit Text

To establish the directionality of the comparison, the full commit ID of the after commit.

Destructuring the Response

data GetCommentsForComparedCommitResponse

See: newGetCommentsForComparedCommitResponse smart constructor.

Constructors

GetCommentsForComparedCommitResponse'

Fields

commentsForComparedCommitData :: Maybe [CommentsForComparedCommit]

A list of comment objects on the compared commit.

nextToken :: Maybe Text

An enumeration token that can be used in a request to return the next batch of the results.

httpStatus :: Int

The response's http status code.

newGetCommentsForComparedCommitResponse

Arguments

:: Int
$sel:httpStatus:GetCommentsForComparedCommitResponse'

-> GetCommentsForComparedCommitResponse

Create a value of GetCommentsForComparedCommitResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:commentsForComparedCommitData:GetCommentsForComparedCommitResponse', getCommentsForComparedCommitResponse_commentsForComparedCommitData - A list of comment objects on the compared commit.

$sel:nextToken:GetCommentsForComparedCommit', getCommentsForComparedCommitResponse_nextToken - An enumeration token that can be used in a request to return the next batch of the results.

$sel:httpStatus:GetCommentsForComparedCommitResponse', getCommentsForComparedCommitResponse_httpStatus - The response's http status code.

Response Lenses

getCommentsForComparedCommitResponse_commentsForComparedCommitData :: Lens' GetCommentsForComparedCommitResponse (Maybe [CommentsForComparedCommit])

A list of comment objects on the compared commit.

getCommentsForComparedCommitResponse_nextToken :: Lens' GetCommentsForComparedCommitResponse (Maybe Text)

An enumeration token that can be used in a request to return the next batch of the results.

getCommentsForComparedCommitResponse_httpStatus :: Lens' GetCommentsForComparedCommitResponse Int

The response's http status code.