Network.AWS.CodeCommit.GetCommentsForPullRequest

Description

Returns comments made on a pull request.

Reaction counts might include numbers from user identities who were deleted after the reaction was made. For a count of reactions from active identities, use GetCommentReactions.

This operation returns paginated results.

Creating a Request

data GetCommentsForPullRequest

See: newGetCommentsForPullRequest smart constructor.

Constructors

GetCommentsForPullRequest'

Fields

nextToken :: Maybe Text

An enumeration token that, when provided in a request, returns the next batch of the results.

maxResults :: Maybe Int

A non-zero, non-negative integer used to limit the number of returned results. The default is 100 comments. You can return up to 500 comments with a single request.

repositoryName :: Maybe Text

The name of the repository that contains the pull request.

beforeCommitId :: Maybe Text

The full commit ID of the commit in the destination branch that was the tip of the branch at the time the pull request was created.

afterCommitId :: Maybe Text

The full commit ID of the commit in the source branch that was the tip of the branch at the time the comment was made.

pullRequestId :: Text

The system-generated ID of the pull request. To get this ID, use ListPullRequests.

newGetCommentsForPullRequest

Arguments

:: Text
$sel:pullRequestId:GetCommentsForPullRequest'

-> GetCommentsForPullRequest

Create a value of GetCommentsForPullRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetCommentsForPullRequest', getCommentsForPullRequest_nextToken - An enumeration token that, when provided in a request, returns the next batch of the results.

$sel:maxResults:GetCommentsForPullRequest', getCommentsForPullRequest_maxResults - A non-zero, non-negative integer used to limit the number of returned results. The default is 100 comments. You can return up to 500 comments with a single request.

$sel:repositoryName:GetCommentsForPullRequest', getCommentsForPullRequest_repositoryName - The name of the repository that contains the pull request.

$sel:beforeCommitId:GetCommentsForPullRequest', getCommentsForPullRequest_beforeCommitId - The full commit ID of the commit in the destination branch that was the tip of the branch at the time the pull request was created.

$sel:afterCommitId:GetCommentsForPullRequest', getCommentsForPullRequest_afterCommitId - The full commit ID of the commit in the source branch that was the tip of the branch at the time the comment was made.

$sel:pullRequestId:GetCommentsForPullRequest', getCommentsForPullRequest_pullRequestId - The system-generated ID of the pull request. To get this ID, use ListPullRequests.

Request Lenses

getCommentsForPullRequest_nextToken :: Lens' GetCommentsForPullRequest (Maybe Text)

An enumeration token that, when provided in a request, returns the next batch of the results.

getCommentsForPullRequest_maxResults :: Lens' GetCommentsForPullRequest (Maybe Int)

A non-zero, non-negative integer used to limit the number of returned results. The default is 100 comments. You can return up to 500 comments with a single request.

getCommentsForPullRequest_repositoryName :: Lens' GetCommentsForPullRequest (Maybe Text)

The name of the repository that contains the pull request.

getCommentsForPullRequest_beforeCommitId :: Lens' GetCommentsForPullRequest (Maybe Text)

The full commit ID of the commit in the destination branch that was the tip of the branch at the time the pull request was created.

getCommentsForPullRequest_afterCommitId :: Lens' GetCommentsForPullRequest (Maybe Text)

The full commit ID of the commit in the source branch that was the tip of the branch at the time the comment was made.

getCommentsForPullRequest_pullRequestId :: Lens' GetCommentsForPullRequest Text

The system-generated ID of the pull request. To get this ID, use ListPullRequests.

Destructuring the Response

data GetCommentsForPullRequestResponse

See: newGetCommentsForPullRequestResponse smart constructor.

Constructors

GetCommentsForPullRequestResponse'

Fields

commentsForPullRequestData :: Maybe [CommentsForPullRequest]

An array of comment objects on the pull request.

nextToken :: Maybe Text

An enumeration token that can be used in a request to return the next batch of the results.

httpStatus :: Int

The response's http status code.

newGetCommentsForPullRequestResponse

Arguments

:: Int
$sel:httpStatus:GetCommentsForPullRequestResponse'

-> GetCommentsForPullRequestResponse

Create a value of GetCommentsForPullRequestResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:commentsForPullRequestData:GetCommentsForPullRequestResponse', getCommentsForPullRequestResponse_commentsForPullRequestData - An array of comment objects on the pull request.

$sel:nextToken:GetCommentsForPullRequest', getCommentsForPullRequestResponse_nextToken - An enumeration token that can be used in a request to return the next batch of the results.

$sel:httpStatus:GetCommentsForPullRequestResponse', getCommentsForPullRequestResponse_httpStatus - The response's http status code.

Response Lenses

getCommentsForPullRequestResponse_commentsForPullRequestData :: Lens' GetCommentsForPullRequestResponse (Maybe [CommentsForPullRequest])

An array of comment objects on the pull request.

getCommentsForPullRequestResponse_nextToken :: Lens' GetCommentsForPullRequestResponse (Maybe Text)

An enumeration token that can be used in a request to return the next batch of the results.

getCommentsForPullRequestResponse_httpStatus :: Lens' GetCommentsForPullRequestResponse Int

The response's http status code.