Network.AWS.CodeCommit.GetDifferences

Description

Returns information about the differences in a valid commit specifier (such as a branch, tag, HEAD, commit ID, or other fully qualified reference). Results can be limited to a specified path.

This operation returns paginated results.

Creating a Request

data GetDifferences

See: newGetDifferences smart constructor.

Constructors

GetDifferences'

Fields

beforeCommitSpecifier :: Maybe Text

The branch, tag, HEAD, or other fully qualified reference used to identify a commit (for example, the full commit ID). Optional. If not specified, all changes before the afterCommitSpecifier value are shown. If you do not use beforeCommitSpecifier in your request, consider limiting the results with maxResults.

nextToken :: Maybe Text

An enumeration token that, when provided in a request, returns the next batch of the results.

maxResults :: Maybe Int

A non-zero, non-negative integer used to limit the number of returned results.

beforePath :: Maybe Text

The file path in which to check for differences. Limits the results to this path. Can also be used to specify the previous name of a directory or folder. If beforePath and afterPath are not specified, differences are shown for all paths.

afterPath :: Maybe Text

The file path in which to check differences. Limits the results to this path. Can also be used to specify the changed name of a directory or folder, if it has changed. If not specified, differences are shown for all paths.

repositoryName :: Text

The name of the repository where you want to get differences.

afterCommitSpecifier :: Text

The branch, tag, HEAD, or other fully qualified reference used to identify a commit.

newGetDifferences

Arguments

:: Text
$sel:repositoryName:GetDifferences'

-> Text
$sel:afterCommitSpecifier:GetDifferences'

-> GetDifferences

Create a value of GetDifferences with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:beforeCommitSpecifier:GetDifferences', getDifferences_beforeCommitSpecifier - The branch, tag, HEAD, or other fully qualified reference used to identify a commit (for example, the full commit ID). Optional. If not specified, all changes before the afterCommitSpecifier value are shown. If you do not use beforeCommitSpecifier in your request, consider limiting the results with maxResults.

$sel:nextToken:GetDifferences', getDifferences_nextToken - An enumeration token that, when provided in a request, returns the next batch of the results.

$sel:maxResults:GetDifferences', getDifferences_maxResults - A non-zero, non-negative integer used to limit the number of returned results.

$sel:beforePath:GetDifferences', getDifferences_beforePath - The file path in which to check for differences. Limits the results to this path. Can also be used to specify the previous name of a directory or folder. If beforePath and afterPath are not specified, differences are shown for all paths.

$sel:afterPath:GetDifferences', getDifferences_afterPath - The file path in which to check differences. Limits the results to this path. Can also be used to specify the changed name of a directory or folder, if it has changed. If not specified, differences are shown for all paths.

$sel:repositoryName:GetDifferences', getDifferences_repositoryName - The name of the repository where you want to get differences.

$sel:afterCommitSpecifier:GetDifferences', getDifferences_afterCommitSpecifier - The branch, tag, HEAD, or other fully qualified reference used to identify a commit.

Request Lenses

getDifferences_beforeCommitSpecifier :: Lens' GetDifferences (Maybe Text)

The branch, tag, HEAD, or other fully qualified reference used to identify a commit (for example, the full commit ID). Optional. If not specified, all changes before the afterCommitSpecifier value are shown. If you do not use beforeCommitSpecifier in your request, consider limiting the results with maxResults.

getDifferences_nextToken :: Lens' GetDifferences (Maybe Text)

An enumeration token that, when provided in a request, returns the next batch of the results.

getDifferences_maxResults :: Lens' GetDifferences (Maybe Int)

A non-zero, non-negative integer used to limit the number of returned results.

getDifferences_beforePath :: Lens' GetDifferences (Maybe Text)

The file path in which to check for differences. Limits the results to this path. Can also be used to specify the previous name of a directory or folder. If beforePath and afterPath are not specified, differences are shown for all paths.

getDifferences_afterPath :: Lens' GetDifferences (Maybe Text)

The file path in which to check differences. Limits the results to this path. Can also be used to specify the changed name of a directory or folder, if it has changed. If not specified, differences are shown for all paths.

getDifferences_repositoryName :: Lens' GetDifferences Text

The name of the repository where you want to get differences.

getDifferences_afterCommitSpecifier :: Lens' GetDifferences Text

The branch, tag, HEAD, or other fully qualified reference used to identify a commit.

Destructuring the Response

data GetDifferencesResponse

See: newGetDifferencesResponse smart constructor.

Constructors

GetDifferencesResponse'

Fields

differences :: Maybe [Difference]

A data type object that contains information about the differences, including whether the difference is added, modified, or deleted (A, D, M).

nextToken :: Maybe Text

An enumeration token that can be used in a request to return the next batch of the results.

httpStatus :: Int

The response's http status code.

newGetDifferencesResponse

Arguments

:: Int
$sel:httpStatus:GetDifferencesResponse'

-> GetDifferencesResponse

Create a value of GetDifferencesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:differences:GetDifferencesResponse', getDifferencesResponse_differences - A data type object that contains information about the differences, including whether the difference is added, modified, or deleted (A, D, M).

$sel:nextToken:GetDifferences', getDifferencesResponse_nextToken - An enumeration token that can be used in a request to return the next batch of the results.

$sel:httpStatus:GetDifferencesResponse', getDifferencesResponse_httpStatus - The response's http status code.

Response Lenses

getDifferencesResponse_differences :: Lens' GetDifferencesResponse (Maybe [Difference])

A data type object that contains information about the differences, including whether the difference is added, modified, or deleted (A, D, M).

getDifferencesResponse_nextToken :: Lens' GetDifferencesResponse (Maybe Text)

An enumeration token that can be used in a request to return the next batch of the results.

getDifferencesResponse_httpStatus :: Lens' GetDifferencesResponse Int

The response's http status code.