Network.AWS.CodeCommit.GetFolder

Description

Returns the contents of a specified folder in a repository.

Creating a Request

data GetFolder

See: newGetFolder smart constructor.

Constructors

GetFolder'

Fields

commitSpecifier :: Maybe Text

A fully qualified reference used to identify a commit that contains the version of the folder's content to return. A fully qualified reference can be a commit ID, branch name, tag, or reference such as HEAD. If no specifier is provided, the folder content is returned as it exists in the HEAD commit.

repositoryName :: Text

The name of the repository.

folderPath :: Text

The fully qualified path to the folder whose contents are returned, including the folder name. For example, /examples is a fully-qualified path to a folder named examples that was created off of the root directory (/) of a repository.

newGetFolder

Arguments

:: Text
$sel:repositoryName:GetFolder'

-> Text
$sel:folderPath:GetFolder'

-> GetFolder

Create a value of GetFolder with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:commitSpecifier:GetFolder', getFolder_commitSpecifier - A fully qualified reference used to identify a commit that contains the version of the folder's content to return. A fully qualified reference can be a commit ID, branch name, tag, or reference such as HEAD. If no specifier is provided, the folder content is returned as it exists in the HEAD commit.

$sel:repositoryName:GetFolder', getFolder_repositoryName - The name of the repository.

$sel:folderPath:GetFolder', getFolder_folderPath - The fully qualified path to the folder whose contents are returned, including the folder name. For example, /examples is a fully-qualified path to a folder named examples that was created off of the root directory (/) of a repository.

Request Lenses

getFolder_commitSpecifier :: Lens' GetFolder (Maybe Text)

A fully qualified reference used to identify a commit that contains the version of the folder's content to return. A fully qualified reference can be a commit ID, branch name, tag, or reference such as HEAD. If no specifier is provided, the folder content is returned as it exists in the HEAD commit.

getFolder_repositoryName :: Lens' GetFolder Text

The name of the repository.

getFolder_folderPath :: Lens' GetFolder Text

The fully qualified path to the folder whose contents are returned, including the folder name. For example, /examples is a fully-qualified path to a folder named examples that was created off of the root directory (/) of a repository.

Destructuring the Response

data GetFolderResponse

See: newGetFolderResponse smart constructor.

Constructors

GetFolderResponse'

Fields

subModules :: Maybe [SubModule]

The list of submodules in the specified folder, if any.

treeId :: Maybe Text

The full SHA-1 pointer of the tree information for the commit that contains the folder.

subFolders :: Maybe [Folder]

The list of folders that exist under the specified folder, if any.

symbolicLinks :: Maybe [SymbolicLink]

The list of symbolic links to other files and folders in the specified folder, if any.

files :: Maybe [File]

The list of files in the specified folder, if any.

httpStatus :: Int

The response's http status code.

commitId :: Text

The full commit ID used as a reference for the returned version of the folder content.

folderPath :: Text

The fully qualified path of the folder whose contents are returned.

newGetFolderResponse

Arguments

:: Int
$sel:httpStatus:GetFolderResponse'

-> Text
$sel:commitId:GetFolderResponse'

-> Text
$sel:folderPath:GetFolder'

-> GetFolderResponse

Create a value of GetFolderResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:subModules:GetFolderResponse', getFolderResponse_subModules - The list of submodules in the specified folder, if any.

$sel:treeId:GetFolderResponse', getFolderResponse_treeId - The full SHA-1 pointer of the tree information for the commit that contains the folder.

$sel:subFolders:GetFolderResponse', getFolderResponse_subFolders - The list of folders that exist under the specified folder, if any.

$sel:symbolicLinks:GetFolderResponse', getFolderResponse_symbolicLinks - The list of symbolic links to other files and folders in the specified folder, if any.

$sel:files:GetFolderResponse', getFolderResponse_files - The list of files in the specified folder, if any.

$sel:httpStatus:GetFolderResponse', getFolderResponse_httpStatus - The response's http status code.

$sel:commitId:GetFolderResponse', getFolderResponse_commitId - The full commit ID used as a reference for the returned version of the folder content.

$sel:folderPath:GetFolder', getFolderResponse_folderPath - The fully qualified path of the folder whose contents are returned.

Response Lenses

getFolderResponse_subModules :: Lens' GetFolderResponse (Maybe [SubModule])

The list of submodules in the specified folder, if any.

getFolderResponse_treeId :: Lens' GetFolderResponse (Maybe Text)

The full SHA-1 pointer of the tree information for the commit that contains the folder.

getFolderResponse_subFolders :: Lens' GetFolderResponse (Maybe [Folder])

The list of folders that exist under the specified folder, if any.

getFolderResponse_symbolicLinks :: Lens' GetFolderResponse (Maybe [SymbolicLink])

The list of symbolic links to other files and folders in the specified folder, if any.

getFolderResponse_files :: Lens' GetFolderResponse (Maybe [File])

The list of files in the specified folder, if any.

getFolderResponse_httpStatus :: Lens' GetFolderResponse Int

The response's http status code.

getFolderResponse_commitId :: Lens' GetFolderResponse Text

The full commit ID used as a reference for the returned version of the folder content.

getFolderResponse_folderPath :: Lens' GetFolderResponse Text

The fully qualified path of the folder whose contents are returned.