Network.AWS.CodeCommit.GetMergeCommit

Description

Returns information about a specified merge commit.

Creating a Request

data GetMergeCommit

See: newGetMergeCommit smart constructor.

Constructors

GetMergeCommit'

Fields

conflictDetailLevel :: Maybe ConflictDetailLevelTypeEnum

The level of conflict detail to use. If unspecified, the default FILE_LEVEL is used, which returns a not-mergeable result if the same file has differences in both branches. If LINE_LEVEL is specified, a conflict is considered not mergeable if the same file in both branches has differences on the same line.

conflictResolutionStrategy :: Maybe ConflictResolutionStrategyTypeEnum

Specifies which branch to use when resolving conflicts, or whether to attempt automatically merging two versions of a file. The default is NONE, which requires any conflicts to be resolved manually before the merge operation is successful.

repositoryName :: Text

The name of the repository that contains the merge commit about which you want to get information.

sourceCommitSpecifier :: Text

The branch, tag, HEAD, or other fully qualified reference used to identify a commit (for example, a branch name or a full commit ID).

destinationCommitSpecifier :: Text

The branch, tag, HEAD, or other fully qualified reference used to identify a commit (for example, a branch name or a full commit ID).

newGetMergeCommit

Arguments

:: Text
$sel:repositoryName:GetMergeCommit'

-> Text
$sel:sourceCommitSpecifier:GetMergeCommit'

-> Text
$sel:destinationCommitSpecifier:GetMergeCommit'

-> GetMergeCommit

Create a value of GetMergeCommit with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:conflictDetailLevel:GetMergeCommit', getMergeCommit_conflictDetailLevel - The level of conflict detail to use. If unspecified, the default FILE_LEVEL is used, which returns a not-mergeable result if the same file has differences in both branches. If LINE_LEVEL is specified, a conflict is considered not mergeable if the same file in both branches has differences on the same line.

$sel:conflictResolutionStrategy:GetMergeCommit', getMergeCommit_conflictResolutionStrategy - Specifies which branch to use when resolving conflicts, or whether to attempt automatically merging two versions of a file. The default is NONE, which requires any conflicts to be resolved manually before the merge operation is successful.

$sel:repositoryName:GetMergeCommit', getMergeCommit_repositoryName - The name of the repository that contains the merge commit about which you want to get information.

$sel:sourceCommitSpecifier:GetMergeCommit', getMergeCommit_sourceCommitSpecifier - The branch, tag, HEAD, or other fully qualified reference used to identify a commit (for example, a branch name or a full commit ID).

$sel:destinationCommitSpecifier:GetMergeCommit', getMergeCommit_destinationCommitSpecifier - The branch, tag, HEAD, or other fully qualified reference used to identify a commit (for example, a branch name or a full commit ID).

Request Lenses

getMergeCommit_conflictDetailLevel :: Lens' GetMergeCommit (Maybe ConflictDetailLevelTypeEnum)

The level of conflict detail to use. If unspecified, the default FILE_LEVEL is used, which returns a not-mergeable result if the same file has differences in both branches. If LINE_LEVEL is specified, a conflict is considered not mergeable if the same file in both branches has differences on the same line.

getMergeCommit_conflictResolutionStrategy :: Lens' GetMergeCommit (Maybe ConflictResolutionStrategyTypeEnum)

Specifies which branch to use when resolving conflicts, or whether to attempt automatically merging two versions of a file. The default is NONE, which requires any conflicts to be resolved manually before the merge operation is successful.

getMergeCommit_repositoryName :: Lens' GetMergeCommit Text

The name of the repository that contains the merge commit about which you want to get information.

getMergeCommit_sourceCommitSpecifier :: Lens' GetMergeCommit Text

The branch, tag, HEAD, or other fully qualified reference used to identify a commit (for example, a branch name or a full commit ID).

getMergeCommit_destinationCommitSpecifier :: Lens' GetMergeCommit Text

The branch, tag, HEAD, or other fully qualified reference used to identify a commit (for example, a branch name or a full commit ID).

Destructuring the Response

data GetMergeCommitResponse

See: newGetMergeCommitResponse smart constructor.

Constructors

GetMergeCommitResponse'

Fields

baseCommitId :: Maybe Text

The commit ID of the merge base.

mergedCommitId :: Maybe Text

The commit ID for the merge commit created when the source branch was merged into the destination branch. If the fast-forward merge strategy was used, there is no merge commit.

sourceCommitId :: Maybe Text

The commit ID of the source commit specifier that was used in the merge evaluation.

destinationCommitId :: Maybe Text

The commit ID of the destination commit specifier that was used in the merge evaluation.

httpStatus :: Int

The response's http status code.

newGetMergeCommitResponse

Arguments

:: Int
$sel:httpStatus:GetMergeCommitResponse'

-> GetMergeCommitResponse

Create a value of GetMergeCommitResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:baseCommitId:GetMergeCommitResponse', getMergeCommitResponse_baseCommitId - The commit ID of the merge base.

$sel:mergedCommitId:GetMergeCommitResponse', getMergeCommitResponse_mergedCommitId - The commit ID for the merge commit created when the source branch was merged into the destination branch. If the fast-forward merge strategy was used, there is no merge commit.

$sel:sourceCommitId:GetMergeCommitResponse', getMergeCommitResponse_sourceCommitId - The commit ID of the source commit specifier that was used in the merge evaluation.

$sel:destinationCommitId:GetMergeCommitResponse', getMergeCommitResponse_destinationCommitId - The commit ID of the destination commit specifier that was used in the merge evaluation.

$sel:httpStatus:GetMergeCommitResponse', getMergeCommitResponse_httpStatus - The response's http status code.

Response Lenses

getMergeCommitResponse_baseCommitId :: Lens' GetMergeCommitResponse (Maybe Text)

The commit ID of the merge base.

getMergeCommitResponse_mergedCommitId :: Lens' GetMergeCommitResponse (Maybe Text)

The commit ID for the merge commit created when the source branch was merged into the destination branch. If the fast-forward merge strategy was used, there is no merge commit.

getMergeCommitResponse_sourceCommitId :: Lens' GetMergeCommitResponse (Maybe Text)

The commit ID of the source commit specifier that was used in the merge evaluation.

getMergeCommitResponse_destinationCommitId :: Lens' GetMergeCommitResponse (Maybe Text)

The commit ID of the destination commit specifier that was used in the merge evaluation.

getMergeCommitResponse_httpStatus :: Lens' GetMergeCommitResponse Int

The response's http status code.