Network.AWS.CodeCommit.GetMergeConflicts

Description

Returns information about merge conflicts between the before and after commit IDs for a pull request in a repository.

Creating a Request

data GetMergeConflicts

See: newGetMergeConflicts smart constructor.

Constructors

GetMergeConflicts'

Fields

nextToken :: Maybe Text

An enumeration token that, when provided in a request, returns the next batch of the results.

maxConflictFiles :: Maybe Int

The maximum number of files to include in the output.

conflictDetailLevel :: Maybe ConflictDetailLevelTypeEnum

The level of conflict detail to use. If unspecified, the default FILE_LEVEL is used, which returns a not-mergeable result if the same file has differences in both branches. If LINE_LEVEL is specified, a conflict is considered not mergeable if the same file in both branches has differences on the same line.

conflictResolutionStrategy :: Maybe ConflictResolutionStrategyTypeEnum

Specifies which branch to use when resolving conflicts, or whether to attempt automatically merging two versions of a file. The default is NONE, which requires any conflicts to be resolved manually before the merge operation is successful.

repositoryName :: Text

The name of the repository where the pull request was created.

destinationCommitSpecifier :: Text

The branch, tag, HEAD, or other fully qualified reference used to identify a commit (for example, a branch name or a full commit ID).

sourceCommitSpecifier :: Text

The branch, tag, HEAD, or other fully qualified reference used to identify a commit (for example, a branch name or a full commit ID).

mergeOption :: MergeOptionTypeEnum

The merge option or strategy you want to use to merge the code.

newGetMergeConflicts

Arguments

:: Text
$sel:repositoryName:GetMergeConflicts'

-> Text
$sel:destinationCommitSpecifier:GetMergeConflicts'

-> Text
$sel:sourceCommitSpecifier:GetMergeConflicts'

-> MergeOptionTypeEnum
$sel:mergeOption:GetMergeConflicts'

-> GetMergeConflicts

Create a value of GetMergeConflicts with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetMergeConflicts', getMergeConflicts_nextToken - An enumeration token that, when provided in a request, returns the next batch of the results.

$sel:maxConflictFiles:GetMergeConflicts', getMergeConflicts_maxConflictFiles - The maximum number of files to include in the output.

$sel:conflictDetailLevel:GetMergeConflicts', getMergeConflicts_conflictDetailLevel - The level of conflict detail to use. If unspecified, the default FILE_LEVEL is used, which returns a not-mergeable result if the same file has differences in both branches. If LINE_LEVEL is specified, a conflict is considered not mergeable if the same file in both branches has differences on the same line.

$sel:conflictResolutionStrategy:GetMergeConflicts', getMergeConflicts_conflictResolutionStrategy - Specifies which branch to use when resolving conflicts, or whether to attempt automatically merging two versions of a file. The default is NONE, which requires any conflicts to be resolved manually before the merge operation is successful.

$sel:repositoryName:GetMergeConflicts', getMergeConflicts_repositoryName - The name of the repository where the pull request was created.

$sel:destinationCommitSpecifier:GetMergeConflicts', getMergeConflicts_destinationCommitSpecifier - The branch, tag, HEAD, or other fully qualified reference used to identify a commit (for example, a branch name or a full commit ID).

$sel:sourceCommitSpecifier:GetMergeConflicts', getMergeConflicts_sourceCommitSpecifier - The branch, tag, HEAD, or other fully qualified reference used to identify a commit (for example, a branch name or a full commit ID).

$sel:mergeOption:GetMergeConflicts', getMergeConflicts_mergeOption - The merge option or strategy you want to use to merge the code.

Request Lenses

getMergeConflicts_nextToken :: Lens' GetMergeConflicts (Maybe Text)

An enumeration token that, when provided in a request, returns the next batch of the results.

getMergeConflicts_maxConflictFiles :: Lens' GetMergeConflicts (Maybe Int)

The maximum number of files to include in the output.

getMergeConflicts_conflictDetailLevel :: Lens' GetMergeConflicts (Maybe ConflictDetailLevelTypeEnum)

The level of conflict detail to use. If unspecified, the default FILE_LEVEL is used, which returns a not-mergeable result if the same file has differences in both branches. If LINE_LEVEL is specified, a conflict is considered not mergeable if the same file in both branches has differences on the same line.

getMergeConflicts_conflictResolutionStrategy :: Lens' GetMergeConflicts (Maybe ConflictResolutionStrategyTypeEnum)

Specifies which branch to use when resolving conflicts, or whether to attempt automatically merging two versions of a file. The default is NONE, which requires any conflicts to be resolved manually before the merge operation is successful.

getMergeConflicts_repositoryName :: Lens' GetMergeConflicts Text

The name of the repository where the pull request was created.

getMergeConflicts_destinationCommitSpecifier :: Lens' GetMergeConflicts Text

The branch, tag, HEAD, or other fully qualified reference used to identify a commit (for example, a branch name or a full commit ID).

getMergeConflicts_sourceCommitSpecifier :: Lens' GetMergeConflicts Text

The branch, tag, HEAD, or other fully qualified reference used to identify a commit (for example, a branch name or a full commit ID).

getMergeConflicts_mergeOption :: Lens' GetMergeConflicts MergeOptionTypeEnum

The merge option or strategy you want to use to merge the code.

Destructuring the Response

data GetMergeConflictsResponse

See: newGetMergeConflictsResponse smart constructor.

Constructors

GetMergeConflictsResponse'

Fields

nextToken :: Maybe Text

An enumeration token that can be used in a request to return the next batch of the results.

baseCommitId :: Maybe Text

The commit ID of the merge base.

httpStatus :: Int

The response's http status code.

mergeable :: Bool

A Boolean value that indicates whether the code is mergeable by the specified merge option.

destinationCommitId :: Text

The commit ID of the destination commit specifier that was used in the merge evaluation.

sourceCommitId :: Text

The commit ID of the source commit specifier that was used in the merge evaluation.

conflictMetadataList :: [ConflictMetadata]

A list of metadata for any conflicting files. If the specified merge strategy is FAST_FORWARD_MERGE, this list is always empty.

newGetMergeConflictsResponse

Arguments

:: Int
$sel:httpStatus:GetMergeConflictsResponse'

-> Bool
$sel:mergeable:GetMergeConflictsResponse'

-> Text
$sel:destinationCommitId:GetMergeConflictsResponse'

-> Text
$sel:sourceCommitId:GetMergeConflictsResponse'

-> GetMergeConflictsResponse

Create a value of GetMergeConflictsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetMergeConflicts', getMergeConflictsResponse_nextToken - An enumeration token that can be used in a request to return the next batch of the results.

$sel:baseCommitId:GetMergeConflictsResponse', getMergeConflictsResponse_baseCommitId - The commit ID of the merge base.

$sel:httpStatus:GetMergeConflictsResponse', getMergeConflictsResponse_httpStatus - The response's http status code.

$sel:mergeable:GetMergeConflictsResponse', getMergeConflictsResponse_mergeable - A Boolean value that indicates whether the code is mergeable by the specified merge option.

$sel:destinationCommitId:GetMergeConflictsResponse', getMergeConflictsResponse_destinationCommitId - The commit ID of the destination commit specifier that was used in the merge evaluation.

$sel:sourceCommitId:GetMergeConflictsResponse', getMergeConflictsResponse_sourceCommitId - The commit ID of the source commit specifier that was used in the merge evaluation.

$sel:conflictMetadataList:GetMergeConflictsResponse', getMergeConflictsResponse_conflictMetadataList - A list of metadata for any conflicting files. If the specified merge strategy is FAST_FORWARD_MERGE, this list is always empty.

Response Lenses

getMergeConflictsResponse_nextToken :: Lens' GetMergeConflictsResponse (Maybe Text)

An enumeration token that can be used in a request to return the next batch of the results.

getMergeConflictsResponse_baseCommitId :: Lens' GetMergeConflictsResponse (Maybe Text)

The commit ID of the merge base.

getMergeConflictsResponse_httpStatus :: Lens' GetMergeConflictsResponse Int

The response's http status code.

getMergeConflictsResponse_mergeable :: Lens' GetMergeConflictsResponse Bool

A Boolean value that indicates whether the code is mergeable by the specified merge option.

getMergeConflictsResponse_destinationCommitId :: Lens' GetMergeConflictsResponse Text

The commit ID of the destination commit specifier that was used in the merge evaluation.

getMergeConflictsResponse_sourceCommitId :: Lens' GetMergeConflictsResponse Text

The commit ID of the source commit specifier that was used in the merge evaluation.

getMergeConflictsResponse_conflictMetadataList :: Lens' GetMergeConflictsResponse [ConflictMetadata]

A list of metadata for any conflicting files. If the specified merge strategy is FAST_FORWARD_MERGE, this list is always empty.