Network.AWS.CodeCommit.GetMergeOptions

Description

Returns information about the merge options available for merging two specified branches. For details about why a merge option is not available, use GetMergeConflicts or DescribeMergeConflicts.

Creating a Request

data GetMergeOptions

See: newGetMergeOptions smart constructor.

Constructors

GetMergeOptions'

Fields

conflictDetailLevel :: Maybe ConflictDetailLevelTypeEnum

The level of conflict detail to use. If unspecified, the default FILE_LEVEL is used, which returns a not-mergeable result if the same file has differences in both branches. If LINE_LEVEL is specified, a conflict is considered not mergeable if the same file in both branches has differences on the same line.

conflictResolutionStrategy :: Maybe ConflictResolutionStrategyTypeEnum

Specifies which branch to use when resolving conflicts, or whether to attempt automatically merging two versions of a file. The default is NONE, which requires any conflicts to be resolved manually before the merge operation is successful.

repositoryName :: Text

The name of the repository that contains the commits about which you want to get merge options.

sourceCommitSpecifier :: Text

The branch, tag, HEAD, or other fully qualified reference used to identify a commit (for example, a branch name or a full commit ID).

destinationCommitSpecifier :: Text

The branch, tag, HEAD, or other fully qualified reference used to identify a commit (for example, a branch name or a full commit ID).

newGetMergeOptions

Arguments

:: Text
$sel:repositoryName:GetMergeOptions'

-> Text
$sel:sourceCommitSpecifier:GetMergeOptions'

-> Text
$sel:destinationCommitSpecifier:GetMergeOptions'

-> GetMergeOptions

Create a value of GetMergeOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:conflictDetailLevel:GetMergeOptions', getMergeOptions_conflictDetailLevel - The level of conflict detail to use. If unspecified, the default FILE_LEVEL is used, which returns a not-mergeable result if the same file has differences in both branches. If LINE_LEVEL is specified, a conflict is considered not mergeable if the same file in both branches has differences on the same line.

$sel:conflictResolutionStrategy:GetMergeOptions', getMergeOptions_conflictResolutionStrategy - Specifies which branch to use when resolving conflicts, or whether to attempt automatically merging two versions of a file. The default is NONE, which requires any conflicts to be resolved manually before the merge operation is successful.

$sel:repositoryName:GetMergeOptions', getMergeOptions_repositoryName - The name of the repository that contains the commits about which you want to get merge options.

$sel:sourceCommitSpecifier:GetMergeOptions', getMergeOptions_sourceCommitSpecifier - The branch, tag, HEAD, or other fully qualified reference used to identify a commit (for example, a branch name or a full commit ID).

$sel:destinationCommitSpecifier:GetMergeOptions', getMergeOptions_destinationCommitSpecifier - The branch, tag, HEAD, or other fully qualified reference used to identify a commit (for example, a branch name or a full commit ID).

Request Lenses

getMergeOptions_conflictDetailLevel :: Lens' GetMergeOptions (Maybe ConflictDetailLevelTypeEnum)

The level of conflict detail to use. If unspecified, the default FILE_LEVEL is used, which returns a not-mergeable result if the same file has differences in both branches. If LINE_LEVEL is specified, a conflict is considered not mergeable if the same file in both branches has differences on the same line.

getMergeOptions_conflictResolutionStrategy :: Lens' GetMergeOptions (Maybe ConflictResolutionStrategyTypeEnum)

Specifies which branch to use when resolving conflicts, or whether to attempt automatically merging two versions of a file. The default is NONE, which requires any conflicts to be resolved manually before the merge operation is successful.

getMergeOptions_repositoryName :: Lens' GetMergeOptions Text

The name of the repository that contains the commits about which you want to get merge options.

getMergeOptions_sourceCommitSpecifier :: Lens' GetMergeOptions Text

The branch, tag, HEAD, or other fully qualified reference used to identify a commit (for example, a branch name or a full commit ID).

getMergeOptions_destinationCommitSpecifier :: Lens' GetMergeOptions Text

The branch, tag, HEAD, or other fully qualified reference used to identify a commit (for example, a branch name or a full commit ID).

Destructuring the Response

data GetMergeOptionsResponse

See: newGetMergeOptionsResponse smart constructor.

Constructors

GetMergeOptionsResponse'

Fields

httpStatus :: Int

The response's http status code.

mergeOptions :: [MergeOptionTypeEnum]

The merge option or strategy used to merge the code.

sourceCommitId :: Text

The commit ID of the source commit specifier that was used in the merge evaluation.

destinationCommitId :: Text

The commit ID of the destination commit specifier that was used in the merge evaluation.

baseCommitId :: Text

The commit ID of the merge base.

newGetMergeOptionsResponse

Arguments

:: Int
$sel:httpStatus:GetMergeOptionsResponse'

-> Text
$sel:sourceCommitId:GetMergeOptionsResponse'

-> Text
$sel:destinationCommitId:GetMergeOptionsResponse'

-> Text
$sel:baseCommitId:GetMergeOptionsResponse'

-> GetMergeOptionsResponse

Create a value of GetMergeOptionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetMergeOptionsResponse', getMergeOptionsResponse_httpStatus - The response's http status code.

$sel:mergeOptions:GetMergeOptionsResponse', getMergeOptionsResponse_mergeOptions - The merge option or strategy used to merge the code.

$sel:sourceCommitId:GetMergeOptionsResponse', getMergeOptionsResponse_sourceCommitId - The commit ID of the source commit specifier that was used in the merge evaluation.

$sel:destinationCommitId:GetMergeOptionsResponse', getMergeOptionsResponse_destinationCommitId - The commit ID of the destination commit specifier that was used in the merge evaluation.

$sel:baseCommitId:GetMergeOptionsResponse', getMergeOptionsResponse_baseCommitId - The commit ID of the merge base.

Response Lenses

getMergeOptionsResponse_httpStatus :: Lens' GetMergeOptionsResponse Int

The response's http status code.

getMergeOptionsResponse_mergeOptions :: Lens' GetMergeOptionsResponse [MergeOptionTypeEnum]

The merge option or strategy used to merge the code.

getMergeOptionsResponse_sourceCommitId :: Lens' GetMergeOptionsResponse Text

The commit ID of the source commit specifier that was used in the merge evaluation.

getMergeOptionsResponse_destinationCommitId :: Lens' GetMergeOptionsResponse Text

The commit ID of the destination commit specifier that was used in the merge evaluation.

getMergeOptionsResponse_baseCommitId :: Lens' GetMergeOptionsResponse Text

The commit ID of the merge base.