Network.AWS.CodeCommit.GetRepositoryTriggers

Description

Gets information about triggers configured for a repository.

Creating a Request

data GetRepositoryTriggers

Represents the input of a get repository triggers operation.

See: newGetRepositoryTriggers smart constructor.

Constructors

GetRepositoryTriggers'

Fields

repositoryName :: Text

The name of the repository for which the trigger is configured.

newGetRepositoryTriggers

Arguments

:: Text
$sel:repositoryName:GetRepositoryTriggers'

-> GetRepositoryTriggers

Create a value of GetRepositoryTriggers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:repositoryName:GetRepositoryTriggers', getRepositoryTriggers_repositoryName - The name of the repository for which the trigger is configured.

Request Lenses

getRepositoryTriggers_repositoryName :: Lens' GetRepositoryTriggers Text

The name of the repository for which the trigger is configured.

Destructuring the Response

data GetRepositoryTriggersResponse

Represents the output of a get repository triggers operation.

See: newGetRepositoryTriggersResponse smart constructor.

Constructors

GetRepositoryTriggersResponse'

Fields

triggers :: Maybe [RepositoryTrigger]

The JSON block of configuration information for each trigger.

configurationId :: Maybe Text

The system-generated unique ID for the trigger.

httpStatus :: Int

The response's http status code.

newGetRepositoryTriggersResponse

Arguments

:: Int
$sel:httpStatus:GetRepositoryTriggersResponse'

-> GetRepositoryTriggersResponse

Create a value of GetRepositoryTriggersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:triggers:GetRepositoryTriggersResponse', getRepositoryTriggersResponse_triggers - The JSON block of configuration information for each trigger.

$sel:configurationId:GetRepositoryTriggersResponse', getRepositoryTriggersResponse_configurationId - The system-generated unique ID for the trigger.

$sel:httpStatus:GetRepositoryTriggersResponse', getRepositoryTriggersResponse_httpStatus - The response's http status code.

Response Lenses

getRepositoryTriggersResponse_triggers :: Lens' GetRepositoryTriggersResponse (Maybe [RepositoryTrigger])

The JSON block of configuration information for each trigger.

getRepositoryTriggersResponse_configurationId :: Lens' GetRepositoryTriggersResponse (Maybe Text)

The system-generated unique ID for the trigger.

getRepositoryTriggersResponse_httpStatus :: Lens' GetRepositoryTriggersResponse Int

The response's http status code.