Network.AWS.CodeCommit.ListPullRequests

Description

Returns a list of pull requests for a specified repository. The return list can be refined by pull request status or pull request author ARN.

This operation returns paginated results.

Creating a Request

data ListPullRequests

See: newListPullRequests smart constructor.

Constructors

ListPullRequests'

Fields

nextToken :: Maybe Text

An enumeration token that, when provided in a request, returns the next batch of the results.

maxResults :: Maybe Int

A non-zero, non-negative integer used to limit the number of returned results.

pullRequestStatus :: Maybe PullRequestStatusEnum

Optional. The status of the pull request. If used, this refines the results to the pull requests that match the specified status.

authorArn :: Maybe Text

Optional. The Amazon Resource Name (ARN) of the user who created the pull request. If used, this filters the results to pull requests created by that user.

repositoryName :: Text

The name of the repository for which you want to list pull requests.

newListPullRequests

Arguments

:: Text
$sel:repositoryName:ListPullRequests'

-> ListPullRequests

Create a value of ListPullRequests with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListPullRequests', listPullRequests_nextToken - An enumeration token that, when provided in a request, returns the next batch of the results.

$sel:maxResults:ListPullRequests', listPullRequests_maxResults - A non-zero, non-negative integer used to limit the number of returned results.

$sel:pullRequestStatus:ListPullRequests', listPullRequests_pullRequestStatus - Optional. The status of the pull request. If used, this refines the results to the pull requests that match the specified status.

$sel:authorArn:ListPullRequests', listPullRequests_authorArn - Optional. The Amazon Resource Name (ARN) of the user who created the pull request. If used, this filters the results to pull requests created by that user.

$sel:repositoryName:ListPullRequests', listPullRequests_repositoryName - The name of the repository for which you want to list pull requests.

Request Lenses

listPullRequests_nextToken :: Lens' ListPullRequests (Maybe Text)

An enumeration token that, when provided in a request, returns the next batch of the results.

listPullRequests_maxResults :: Lens' ListPullRequests (Maybe Int)

A non-zero, non-negative integer used to limit the number of returned results.

listPullRequests_pullRequestStatus :: Lens' ListPullRequests (Maybe PullRequestStatusEnum)

Optional. The status of the pull request. If used, this refines the results to the pull requests that match the specified status.

listPullRequests_authorArn :: Lens' ListPullRequests (Maybe Text)

Optional. The Amazon Resource Name (ARN) of the user who created the pull request. If used, this filters the results to pull requests created by that user.

listPullRequests_repositoryName :: Lens' ListPullRequests Text

The name of the repository for which you want to list pull requests.

Destructuring the Response

data ListPullRequestsResponse

See: newListPullRequestsResponse smart constructor.

Constructors

ListPullRequestsResponse'

Fields

nextToken :: Maybe Text

An enumeration token that allows the operation to batch the next results of the operation.

httpStatus :: Int

The response's http status code.

pullRequestIds :: [Text]

The system-generated IDs of the pull requests.

newListPullRequestsResponse

Arguments

:: Int
$sel:httpStatus:ListPullRequestsResponse'

-> ListPullRequestsResponse

Create a value of ListPullRequestsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListPullRequests', listPullRequestsResponse_nextToken - An enumeration token that allows the operation to batch the next results of the operation.

$sel:httpStatus:ListPullRequestsResponse', listPullRequestsResponse_httpStatus - The response's http status code.

$sel:pullRequestIds:ListPullRequestsResponse', listPullRequestsResponse_pullRequestIds - The system-generated IDs of the pull requests.

Response Lenses

listPullRequestsResponse_nextToken :: Lens' ListPullRequestsResponse (Maybe Text)

An enumeration token that allows the operation to batch the next results of the operation.

listPullRequestsResponse_httpStatus :: Lens' ListPullRequestsResponse Int

The response's http status code.

listPullRequestsResponse_pullRequestIds :: Lens' ListPullRequestsResponse [Text]

The system-generated IDs of the pull requests.