Network.AWS.CodeCommit.MergeBranchesBySquash

Description

Merges two branches using the squash merge strategy.

Creating a Request

data MergeBranchesBySquash

See: newMergeBranchesBySquash smart constructor.

Constructors

MergeBranchesBySquash'

Fields

authorName :: Maybe Text

The name of the author who created the commit. This information is used as both the author and committer for the commit.

commitMessage :: Maybe Text

The commit message for the merge.

email :: Maybe Text

The email address of the person merging the branches. This information is used in the commit information for the merge.

conflictDetailLevel :: Maybe ConflictDetailLevelTypeEnum

The level of conflict detail to use. If unspecified, the default FILE_LEVEL is used, which returns a not-mergeable result if the same file has differences in both branches. If LINE_LEVEL is specified, a conflict is considered not mergeable if the same file in both branches has differences on the same line.

keepEmptyFolders :: Maybe Bool

If the commit contains deletions, whether to keep a folder or folder structure if the changes leave the folders empty. If this is specified as true, a .gitkeep file is created for empty folders. The default is false.

conflictResolutionStrategy :: Maybe ConflictResolutionStrategyTypeEnum

Specifies which branch to use when resolving conflicts, or whether to attempt automatically merging two versions of a file. The default is NONE, which requires any conflicts to be resolved manually before the merge operation is successful.

conflictResolution :: Maybe ConflictResolution

If AUTOMERGE is the conflict resolution strategy, a list of inputs to use when resolving conflicts during a merge.

targetBranch :: Maybe Text

The branch where the merge is applied.

repositoryName :: Text

The name of the repository where you want to merge two branches.

sourceCommitSpecifier :: Text

The branch, tag, HEAD, or other fully qualified reference used to identify a commit (for example, a branch name or a full commit ID).

destinationCommitSpecifier :: Text

The branch, tag, HEAD, or other fully qualified reference used to identify a commit (for example, a branch name or a full commit ID).

newMergeBranchesBySquash

Arguments

:: Text
$sel:repositoryName:MergeBranchesBySquash'

-> Text
$sel:sourceCommitSpecifier:MergeBranchesBySquash'

-> Text
$sel:destinationCommitSpecifier:MergeBranchesBySquash'

-> MergeBranchesBySquash

Create a value of MergeBranchesBySquash with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:authorName:MergeBranchesBySquash', mergeBranchesBySquash_authorName - The name of the author who created the commit. This information is used as both the author and committer for the commit.

$sel:commitMessage:MergeBranchesBySquash', mergeBranchesBySquash_commitMessage - The commit message for the merge.

$sel:email:MergeBranchesBySquash', mergeBranchesBySquash_email - The email address of the person merging the branches. This information is used in the commit information for the merge.

$sel:conflictDetailLevel:MergeBranchesBySquash', mergeBranchesBySquash_conflictDetailLevel - The level of conflict detail to use. If unspecified, the default FILE_LEVEL is used, which returns a not-mergeable result if the same file has differences in both branches. If LINE_LEVEL is specified, a conflict is considered not mergeable if the same file in both branches has differences on the same line.

$sel:keepEmptyFolders:MergeBranchesBySquash', mergeBranchesBySquash_keepEmptyFolders - If the commit contains deletions, whether to keep a folder or folder structure if the changes leave the folders empty. If this is specified as true, a .gitkeep file is created for empty folders. The default is false.

$sel:conflictResolutionStrategy:MergeBranchesBySquash', mergeBranchesBySquash_conflictResolutionStrategy - Specifies which branch to use when resolving conflicts, or whether to attempt automatically merging two versions of a file. The default is NONE, which requires any conflicts to be resolved manually before the merge operation is successful.

$sel:conflictResolution:MergeBranchesBySquash', mergeBranchesBySquash_conflictResolution - If AUTOMERGE is the conflict resolution strategy, a list of inputs to use when resolving conflicts during a merge.

$sel:targetBranch:MergeBranchesBySquash', mergeBranchesBySquash_targetBranch - The branch where the merge is applied.

$sel:repositoryName:MergeBranchesBySquash', mergeBranchesBySquash_repositoryName - The name of the repository where you want to merge two branches.

$sel:sourceCommitSpecifier:MergeBranchesBySquash', mergeBranchesBySquash_sourceCommitSpecifier - The branch, tag, HEAD, or other fully qualified reference used to identify a commit (for example, a branch name or a full commit ID).

$sel:destinationCommitSpecifier:MergeBranchesBySquash', mergeBranchesBySquash_destinationCommitSpecifier - The branch, tag, HEAD, or other fully qualified reference used to identify a commit (for example, a branch name or a full commit ID).

Request Lenses

mergeBranchesBySquash_authorName :: Lens' MergeBranchesBySquash (Maybe Text)

The name of the author who created the commit. This information is used as both the author and committer for the commit.

mergeBranchesBySquash_commitMessage :: Lens' MergeBranchesBySquash (Maybe Text)

The commit message for the merge.

mergeBranchesBySquash_email :: Lens' MergeBranchesBySquash (Maybe Text)

The email address of the person merging the branches. This information is used in the commit information for the merge.

mergeBranchesBySquash_conflictDetailLevel :: Lens' MergeBranchesBySquash (Maybe ConflictDetailLevelTypeEnum)

The level of conflict detail to use. If unspecified, the default FILE_LEVEL is used, which returns a not-mergeable result if the same file has differences in both branches. If LINE_LEVEL is specified, a conflict is considered not mergeable if the same file in both branches has differences on the same line.

mergeBranchesBySquash_keepEmptyFolders :: Lens' MergeBranchesBySquash (Maybe Bool)

If the commit contains deletions, whether to keep a folder or folder structure if the changes leave the folders empty. If this is specified as true, a .gitkeep file is created for empty folders. The default is false.

mergeBranchesBySquash_conflictResolutionStrategy :: Lens' MergeBranchesBySquash (Maybe ConflictResolutionStrategyTypeEnum)

Specifies which branch to use when resolving conflicts, or whether to attempt automatically merging two versions of a file. The default is NONE, which requires any conflicts to be resolved manually before the merge operation is successful.

mergeBranchesBySquash_conflictResolution :: Lens' MergeBranchesBySquash (Maybe ConflictResolution)

If AUTOMERGE is the conflict resolution strategy, a list of inputs to use when resolving conflicts during a merge.

mergeBranchesBySquash_targetBranch :: Lens' MergeBranchesBySquash (Maybe Text)

The branch where the merge is applied.

mergeBranchesBySquash_repositoryName :: Lens' MergeBranchesBySquash Text

The name of the repository where you want to merge two branches.

mergeBranchesBySquash_sourceCommitSpecifier :: Lens' MergeBranchesBySquash Text

The branch, tag, HEAD, or other fully qualified reference used to identify a commit (for example, a branch name or a full commit ID).

mergeBranchesBySquash_destinationCommitSpecifier :: Lens' MergeBranchesBySquash Text

The branch, tag, HEAD, or other fully qualified reference used to identify a commit (for example, a branch name or a full commit ID).

Destructuring the Response

data MergeBranchesBySquashResponse

See: newMergeBranchesBySquashResponse smart constructor.

Constructors

MergeBranchesBySquashResponse'

Fields

commitId :: Maybe Text

The commit ID of the merge in the destination or target branch.

treeId :: Maybe Text

The tree ID of the merge in the destination or target branch.

httpStatus :: Int

The response's http status code.

newMergeBranchesBySquashResponse

Arguments

:: Int
$sel:httpStatus:MergeBranchesBySquashResponse'

-> MergeBranchesBySquashResponse

Create a value of MergeBranchesBySquashResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:commitId:MergeBranchesBySquashResponse', mergeBranchesBySquashResponse_commitId - The commit ID of the merge in the destination or target branch.

$sel:treeId:MergeBranchesBySquashResponse', mergeBranchesBySquashResponse_treeId - The tree ID of the merge in the destination or target branch.

$sel:httpStatus:MergeBranchesBySquashResponse', mergeBranchesBySquashResponse_httpStatus - The response's http status code.

Response Lenses

mergeBranchesBySquashResponse_commitId :: Lens' MergeBranchesBySquashResponse (Maybe Text)

The commit ID of the merge in the destination or target branch.

mergeBranchesBySquashResponse_treeId :: Lens' MergeBranchesBySquashResponse (Maybe Text)

The tree ID of the merge in the destination or target branch.

mergeBranchesBySquashResponse_httpStatus :: Lens' MergeBranchesBySquashResponse Int

The response's http status code.