Network.AWS.CodeCommit.MergeBranchesByThreeWay

Description

Merges two specified branches using the three-way merge strategy.

Creating a Request

data MergeBranchesByThreeWay

See: newMergeBranchesByThreeWay smart constructor.

Constructors

MergeBranchesByThreeWay'

Fields

email :: Maybe Text

The email address of the person merging the branches. This information is used in the commit information for the merge.

authorName :: Maybe Text

The name of the author who created the commit. This information is used as both the author and committer for the commit.

targetBranch :: Maybe Text

The branch where the merge is applied.

conflictDetailLevel :: Maybe ConflictDetailLevelTypeEnum

The level of conflict detail to use. If unspecified, the default FILE_LEVEL is used, which returns a not-mergeable result if the same file has differences in both branches. If LINE_LEVEL is specified, a conflict is considered not mergeable if the same file in both branches has differences on the same line.

commitMessage :: Maybe Text

The commit message to include in the commit information for the merge.

conflictResolution :: Maybe ConflictResolution

If AUTOMERGE is the conflict resolution strategy, a list of inputs to use when resolving conflicts during a merge.

conflictResolutionStrategy :: Maybe ConflictResolutionStrategyTypeEnum

Specifies which branch to use when resolving conflicts, or whether to attempt automatically merging two versions of a file. The default is NONE, which requires any conflicts to be resolved manually before the merge operation is successful.

keepEmptyFolders :: Maybe Bool

If the commit contains deletions, whether to keep a folder or folder structure if the changes leave the folders empty. If true, a .gitkeep file is created for empty folders. The default is false.

repositoryName :: Text

The name of the repository where you want to merge two branches.

sourceCommitSpecifier :: Text

The branch, tag, HEAD, or other fully qualified reference used to identify a commit (for example, a branch name or a full commit ID).

destinationCommitSpecifier :: Text

The branch, tag, HEAD, or other fully qualified reference used to identify a commit (for example, a branch name or a full commit ID).

newMergeBranchesByThreeWay

Arguments

:: Text
$sel:repositoryName:MergeBranchesByThreeWay'

-> Text
$sel:sourceCommitSpecifier:MergeBranchesByThreeWay'

-> Text
$sel:destinationCommitSpecifier:MergeBranchesByThreeWay'

-> MergeBranchesByThreeWay

Create a value of MergeBranchesByThreeWay with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:email:MergeBranchesByThreeWay', mergeBranchesByThreeWay_email - The email address of the person merging the branches. This information is used in the commit information for the merge.

$sel:authorName:MergeBranchesByThreeWay', mergeBranchesByThreeWay_authorName - The name of the author who created the commit. This information is used as both the author and committer for the commit.

$sel:targetBranch:MergeBranchesByThreeWay', mergeBranchesByThreeWay_targetBranch - The branch where the merge is applied.

$sel:conflictDetailLevel:MergeBranchesByThreeWay', mergeBranchesByThreeWay_conflictDetailLevel - The level of conflict detail to use. If unspecified, the default FILE_LEVEL is used, which returns a not-mergeable result if the same file has differences in both branches. If LINE_LEVEL is specified, a conflict is considered not mergeable if the same file in both branches has differences on the same line.

$sel:commitMessage:MergeBranchesByThreeWay', mergeBranchesByThreeWay_commitMessage - The commit message to include in the commit information for the merge.

$sel:conflictResolution:MergeBranchesByThreeWay', mergeBranchesByThreeWay_conflictResolution - If AUTOMERGE is the conflict resolution strategy, a list of inputs to use when resolving conflicts during a merge.

$sel:conflictResolutionStrategy:MergeBranchesByThreeWay', mergeBranchesByThreeWay_conflictResolutionStrategy - Specifies which branch to use when resolving conflicts, or whether to attempt automatically merging two versions of a file. The default is NONE, which requires any conflicts to be resolved manually before the merge operation is successful.

$sel:keepEmptyFolders:MergeBranchesByThreeWay', mergeBranchesByThreeWay_keepEmptyFolders - If the commit contains deletions, whether to keep a folder or folder structure if the changes leave the folders empty. If true, a .gitkeep file is created for empty folders. The default is false.

$sel:repositoryName:MergeBranchesByThreeWay', mergeBranchesByThreeWay_repositoryName - The name of the repository where you want to merge two branches.

$sel:sourceCommitSpecifier:MergeBranchesByThreeWay', mergeBranchesByThreeWay_sourceCommitSpecifier - The branch, tag, HEAD, or other fully qualified reference used to identify a commit (for example, a branch name or a full commit ID).

$sel:destinationCommitSpecifier:MergeBranchesByThreeWay', mergeBranchesByThreeWay_destinationCommitSpecifier - The branch, tag, HEAD, or other fully qualified reference used to identify a commit (for example, a branch name or a full commit ID).

Request Lenses

mergeBranchesByThreeWay_email :: Lens' MergeBranchesByThreeWay (Maybe Text)

The email address of the person merging the branches. This information is used in the commit information for the merge.

mergeBranchesByThreeWay_authorName :: Lens' MergeBranchesByThreeWay (Maybe Text)

The name of the author who created the commit. This information is used as both the author and committer for the commit.

mergeBranchesByThreeWay_targetBranch :: Lens' MergeBranchesByThreeWay (Maybe Text)

The branch where the merge is applied.

mergeBranchesByThreeWay_conflictDetailLevel :: Lens' MergeBranchesByThreeWay (Maybe ConflictDetailLevelTypeEnum)

The level of conflict detail to use. If unspecified, the default FILE_LEVEL is used, which returns a not-mergeable result if the same file has differences in both branches. If LINE_LEVEL is specified, a conflict is considered not mergeable if the same file in both branches has differences on the same line.

mergeBranchesByThreeWay_commitMessage :: Lens' MergeBranchesByThreeWay (Maybe Text)

The commit message to include in the commit information for the merge.

mergeBranchesByThreeWay_conflictResolution :: Lens' MergeBranchesByThreeWay (Maybe ConflictResolution)

If AUTOMERGE is the conflict resolution strategy, a list of inputs to use when resolving conflicts during a merge.

mergeBranchesByThreeWay_conflictResolutionStrategy :: Lens' MergeBranchesByThreeWay (Maybe ConflictResolutionStrategyTypeEnum)

Specifies which branch to use when resolving conflicts, or whether to attempt automatically merging two versions of a file. The default is NONE, which requires any conflicts to be resolved manually before the merge operation is successful.

mergeBranchesByThreeWay_keepEmptyFolders :: Lens' MergeBranchesByThreeWay (Maybe Bool)

If the commit contains deletions, whether to keep a folder or folder structure if the changes leave the folders empty. If true, a .gitkeep file is created for empty folders. The default is false.

mergeBranchesByThreeWay_repositoryName :: Lens' MergeBranchesByThreeWay Text

The name of the repository where you want to merge two branches.

mergeBranchesByThreeWay_sourceCommitSpecifier :: Lens' MergeBranchesByThreeWay Text

The branch, tag, HEAD, or other fully qualified reference used to identify a commit (for example, a branch name or a full commit ID).

mergeBranchesByThreeWay_destinationCommitSpecifier :: Lens' MergeBranchesByThreeWay Text

The branch, tag, HEAD, or other fully qualified reference used to identify a commit (for example, a branch name or a full commit ID).

Destructuring the Response

data MergeBranchesByThreeWayResponse

See: newMergeBranchesByThreeWayResponse smart constructor.

Constructors

MergeBranchesByThreeWayResponse'

Fields

commitId :: Maybe Text

The commit ID of the merge in the destination or target branch.

treeId :: Maybe Text

The tree ID of the merge in the destination or target branch.

httpStatus :: Int

The response's http status code.

newMergeBranchesByThreeWayResponse

Arguments

:: Int
$sel:httpStatus:MergeBranchesByThreeWayResponse'

-> MergeBranchesByThreeWayResponse

Create a value of MergeBranchesByThreeWayResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:commitId:MergeBranchesByThreeWayResponse', mergeBranchesByThreeWayResponse_commitId - The commit ID of the merge in the destination or target branch.

$sel:treeId:MergeBranchesByThreeWayResponse', mergeBranchesByThreeWayResponse_treeId - The tree ID of the merge in the destination or target branch.

$sel:httpStatus:MergeBranchesByThreeWayResponse', mergeBranchesByThreeWayResponse_httpStatus - The response's http status code.

Response Lenses

mergeBranchesByThreeWayResponse_commitId :: Lens' MergeBranchesByThreeWayResponse (Maybe Text)

The commit ID of the merge in the destination or target branch.

mergeBranchesByThreeWayResponse_treeId :: Lens' MergeBranchesByThreeWayResponse (Maybe Text)

The tree ID of the merge in the destination or target branch.

mergeBranchesByThreeWayResponse_httpStatus :: Lens' MergeBranchesByThreeWayResponse Int

The response's http status code.