Network.AWS.CodeCommit.PostCommentForComparedCommit

Description

Posts a comment on the comparison between two commits.

Creating a Request

data PostCommentForComparedCommit

See: newPostCommentForComparedCommit smart constructor.

Constructors

PostCommentForComparedCommit'

Fields

location :: Maybe Location

The location of the comparison where you want to comment.

beforeCommitId :: Maybe Text

To establish the directionality of the comparison, the full commit ID of the before commit. Required for commenting on any commit unless that commit is the initial commit.

clientRequestToken :: Maybe Text

A unique, client-generated idempotency token that, when provided in a request, ensures the request cannot be repeated with a changed parameter. If a request is received with the same parameters and a token is included, the request returns information about the initial request that used that token.

repositoryName :: Text

The name of the repository where you want to post a comment on the comparison between commits.

afterCommitId :: Text

To establish the directionality of the comparison, the full commit ID of the after commit.

content :: Text

The content of the comment you want to make.

newPostCommentForComparedCommit

Arguments

:: Text
$sel:repositoryName:PostCommentForComparedCommit'

-> Text
$sel:afterCommitId:PostCommentForComparedCommit'

-> Text
$sel:content:PostCommentForComparedCommit'

-> PostCommentForComparedCommit

Create a value of PostCommentForComparedCommit with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:location:PostCommentForComparedCommit', postCommentForComparedCommit_location - The location of the comparison where you want to comment.

$sel:beforeCommitId:PostCommentForComparedCommit', postCommentForComparedCommit_beforeCommitId - To establish the directionality of the comparison, the full commit ID of the before commit. Required for commenting on any commit unless that commit is the initial commit.

$sel:clientRequestToken:PostCommentForComparedCommit', postCommentForComparedCommit_clientRequestToken - A unique, client-generated idempotency token that, when provided in a request, ensures the request cannot be repeated with a changed parameter. If a request is received with the same parameters and a token is included, the request returns information about the initial request that used that token.

$sel:repositoryName:PostCommentForComparedCommit', postCommentForComparedCommit_repositoryName - The name of the repository where you want to post a comment on the comparison between commits.

$sel:afterCommitId:PostCommentForComparedCommit', postCommentForComparedCommit_afterCommitId - To establish the directionality of the comparison, the full commit ID of the after commit.

$sel:content:PostCommentForComparedCommit', postCommentForComparedCommit_content - The content of the comment you want to make.

Request Lenses

postCommentForComparedCommit_location :: Lens' PostCommentForComparedCommit (Maybe Location)

The location of the comparison where you want to comment.

postCommentForComparedCommit_beforeCommitId :: Lens' PostCommentForComparedCommit (Maybe Text)

To establish the directionality of the comparison, the full commit ID of the before commit. Required for commenting on any commit unless that commit is the initial commit.

postCommentForComparedCommit_clientRequestToken :: Lens' PostCommentForComparedCommit (Maybe Text)

A unique, client-generated idempotency token that, when provided in a request, ensures the request cannot be repeated with a changed parameter. If a request is received with the same parameters and a token is included, the request returns information about the initial request that used that token.

postCommentForComparedCommit_repositoryName :: Lens' PostCommentForComparedCommit Text

The name of the repository where you want to post a comment on the comparison between commits.

postCommentForComparedCommit_afterCommitId :: Lens' PostCommentForComparedCommit Text

To establish the directionality of the comparison, the full commit ID of the after commit.

postCommentForComparedCommit_content :: Lens' PostCommentForComparedCommit Text

The content of the comment you want to make.

Destructuring the Response

data PostCommentForComparedCommitResponse

See: newPostCommentForComparedCommitResponse smart constructor.

Constructors

PostCommentForComparedCommitResponse'

Fields

beforeBlobId :: Maybe Text

In the directionality you established, the blob ID of the before blob.

location :: Maybe Location

The location of the comment in the comparison between the two commits.

afterCommitId :: Maybe Text

In the directionality you established, the full commit ID of the after commit.

afterBlobId :: Maybe Text

In the directionality you established, the blob ID of the after blob.

beforeCommitId :: Maybe Text

In the directionality you established, the full commit ID of the before commit.

repositoryName :: Maybe Text

The name of the repository where you posted a comment on the comparison between commits.

comment :: Maybe Comment

The content of the comment you posted.

httpStatus :: Int

The response's http status code.

newPostCommentForComparedCommitResponse

Arguments

:: Int
$sel:httpStatus:PostCommentForComparedCommitResponse'

-> PostCommentForComparedCommitResponse

Create a value of PostCommentForComparedCommitResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:beforeBlobId:PostCommentForComparedCommitResponse', postCommentForComparedCommitResponse_beforeBlobId - In the directionality you established, the blob ID of the before blob.

$sel:location:PostCommentForComparedCommit', postCommentForComparedCommitResponse_location - The location of the comment in the comparison between the two commits.

$sel:afterCommitId:PostCommentForComparedCommit', postCommentForComparedCommitResponse_afterCommitId - In the directionality you established, the full commit ID of the after commit.

$sel:afterBlobId:PostCommentForComparedCommitResponse', postCommentForComparedCommitResponse_afterBlobId - In the directionality you established, the blob ID of the after blob.

$sel:beforeCommitId:PostCommentForComparedCommit', postCommentForComparedCommitResponse_beforeCommitId - In the directionality you established, the full commit ID of the before commit.

$sel:repositoryName:PostCommentForComparedCommit', postCommentForComparedCommitResponse_repositoryName - The name of the repository where you posted a comment on the comparison between commits.

$sel:comment:PostCommentForComparedCommitResponse', postCommentForComparedCommitResponse_comment - The content of the comment you posted.

$sel:httpStatus:PostCommentForComparedCommitResponse', postCommentForComparedCommitResponse_httpStatus - The response's http status code.

Response Lenses

postCommentForComparedCommitResponse_beforeBlobId :: Lens' PostCommentForComparedCommitResponse (Maybe Text)

In the directionality you established, the blob ID of the before blob.

postCommentForComparedCommitResponse_location :: Lens' PostCommentForComparedCommitResponse (Maybe Location)

The location of the comment in the comparison between the two commits.

postCommentForComparedCommitResponse_afterCommitId :: Lens' PostCommentForComparedCommitResponse (Maybe Text)

In the directionality you established, the full commit ID of the after commit.

postCommentForComparedCommitResponse_afterBlobId :: Lens' PostCommentForComparedCommitResponse (Maybe Text)

In the directionality you established, the blob ID of the after blob.

postCommentForComparedCommitResponse_beforeCommitId :: Lens' PostCommentForComparedCommitResponse (Maybe Text)

In the directionality you established, the full commit ID of the before commit.

postCommentForComparedCommitResponse_repositoryName :: Lens' PostCommentForComparedCommitResponse (Maybe Text)

The name of the repository where you posted a comment on the comparison between commits.

postCommentForComparedCommitResponse_comment :: Lens' PostCommentForComparedCommitResponse (Maybe Comment)

The content of the comment you posted.

postCommentForComparedCommitResponse_httpStatus :: Lens' PostCommentForComparedCommitResponse Int

The response's http status code.