Network.AWS.CodeCommit.PutFile

Description

Adds or updates a file in a branch in an AWS CodeCommit repository, and generates a commit for the addition in the specified branch.

Creating a Request

data PutFile

See: newPutFile smart constructor.

Constructors

PutFile'

Fields

parentCommitId :: Maybe Text

The full commit ID of the head commit in the branch where you want to add or update the file. If this is an empty repository, no commit ID is required. If this is not an empty repository, a commit ID is required.

The commit ID must match the ID of the head commit at the time of the operation. Otherwise, an error occurs, and the file is not added or updated.

commitMessage :: Maybe Text

A message about why this file was added or updated. Although it is optional, a message makes the commit history for your repository more useful.

name :: Maybe Text

The name of the person adding or updating the file. Although it is optional, a name makes the commit history for your repository more useful.

email :: Maybe Text

An email address for the person adding or updating the file.

fileMode :: Maybe FileModeTypeEnum

The file mode permissions of the blob. Valid file mode permissions are listed here.

repositoryName :: Text

The name of the repository where you want to add or update the file.

branchName :: Text

The name of the branch where you want to add or update the file. If this is an empty repository, this branch is created.

fileContent :: Base64

The content of the file, in binary object format.

filePath :: Text

The name of the file you want to add or update, including the relative path to the file in the repository.

If the path does not currently exist in the repository, the path is created as part of adding the file.

newPutFile

Arguments

:: Text
$sel:repositoryName:PutFile'

-> Text
$sel:branchName:PutFile'

-> ByteString
$sel:fileContent:PutFile'

-> Text
$sel:filePath:PutFile'

-> PutFile

Create a value of PutFile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parentCommitId:PutFile', putFile_parentCommitId - The full commit ID of the head commit in the branch where you want to add or update the file. If this is an empty repository, no commit ID is required. If this is not an empty repository, a commit ID is required.

The commit ID must match the ID of the head commit at the time of the operation. Otherwise, an error occurs, and the file is not added or updated.

$sel:commitMessage:PutFile', putFile_commitMessage - A message about why this file was added or updated. Although it is optional, a message makes the commit history for your repository more useful.

$sel:name:PutFile', putFile_name - The name of the person adding or updating the file. Although it is optional, a name makes the commit history for your repository more useful.

$sel:email:PutFile', putFile_email - An email address for the person adding or updating the file.

$sel:fileMode:PutFile', putFile_fileMode - The file mode permissions of the blob. Valid file mode permissions are listed here.

$sel:repositoryName:PutFile', putFile_repositoryName - The name of the repository where you want to add or update the file.

$sel:branchName:PutFile', putFile_branchName - The name of the branch where you want to add or update the file. If this is an empty repository, this branch is created.

$sel:fileContent:PutFile', putFile_fileContent - The content of the file, in binary object format.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:filePath:PutFile', putFile_filePath - The name of the file you want to add or update, including the relative path to the file in the repository.

If the path does not currently exist in the repository, the path is created as part of adding the file.

Request Lenses

putFile_parentCommitId :: Lens' PutFile (Maybe Text)

The full commit ID of the head commit in the branch where you want to add or update the file. If this is an empty repository, no commit ID is required. If this is not an empty repository, a commit ID is required.

The commit ID must match the ID of the head commit at the time of the operation. Otherwise, an error occurs, and the file is not added or updated.

putFile_commitMessage :: Lens' PutFile (Maybe Text)

A message about why this file was added or updated. Although it is optional, a message makes the commit history for your repository more useful.

putFile_name :: Lens' PutFile (Maybe Text)

The name of the person adding or updating the file. Although it is optional, a name makes the commit history for your repository more useful.

putFile_email :: Lens' PutFile (Maybe Text)

An email address for the person adding or updating the file.

putFile_fileMode :: Lens' PutFile (Maybe FileModeTypeEnum)

The file mode permissions of the blob. Valid file mode permissions are listed here.

putFile_repositoryName :: Lens' PutFile Text

The name of the repository where you want to add or update the file.

putFile_branchName :: Lens' PutFile Text

The name of the branch where you want to add or update the file. If this is an empty repository, this branch is created.

putFile_fileContent :: Lens' PutFile ByteString

The content of the file, in binary object format.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

putFile_filePath :: Lens' PutFile Text

The name of the file you want to add or update, including the relative path to the file in the repository.

If the path does not currently exist in the repository, the path is created as part of adding the file.

Destructuring the Response

data PutFileResponse

See: newPutFileResponse smart constructor.

Constructors

PutFileResponse'

Fields

httpStatus :: Int

The response's http status code.

commitId :: Text

The full SHA ID of the commit that contains this file change.

blobId :: Text

The ID of the blob, which is its SHA-1 pointer.

treeId :: Text

The full SHA-1 pointer of the tree information for the commit that contains this file change.

newPutFileResponse

Arguments

:: Int
$sel:httpStatus:PutFileResponse'

-> Text
$sel:commitId:PutFileResponse'

-> Text
$sel:blobId:PutFileResponse'

-> Text
$sel:treeId:PutFileResponse'

-> PutFileResponse

Create a value of PutFileResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:PutFileResponse', putFileResponse_httpStatus - The response's http status code.

$sel:commitId:PutFileResponse', putFileResponse_commitId - The full SHA ID of the commit that contains this file change.

$sel:blobId:PutFileResponse', putFileResponse_blobId - The ID of the blob, which is its SHA-1 pointer.

$sel:treeId:PutFileResponse', putFileResponse_treeId - The full SHA-1 pointer of the tree information for the commit that contains this file change.

Response Lenses

putFileResponse_httpStatus :: Lens' PutFileResponse Int

The response's http status code.

putFileResponse_commitId :: Lens' PutFileResponse Text

The full SHA ID of the commit that contains this file change.

putFileResponse_blobId :: Lens' PutFileResponse Text

The ID of the blob, which is its SHA-1 pointer.

putFileResponse_treeId :: Lens' PutFileResponse Text

The full SHA-1 pointer of the tree information for the commit that contains this file change.