Network.AWS.CodeCommit.TestRepositoryTriggers

Description

Tests the functionality of repository triggers by sending information to the trigger target. If real data is available in the repository, the test sends data from the last commit. If no data is available, sample data is generated.

Creating a Request

data TestRepositoryTriggers

Represents the input of a test repository triggers operation.

See: newTestRepositoryTriggers smart constructor.

Constructors

TestRepositoryTriggers'

Fields

repositoryName :: Text

The name of the repository in which to test the triggers.

triggers :: [RepositoryTrigger]

The list of triggers to test.

newTestRepositoryTriggers

Arguments

:: Text
$sel:repositoryName:TestRepositoryTriggers'

-> TestRepositoryTriggers

Create a value of TestRepositoryTriggers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:repositoryName:TestRepositoryTriggers', testRepositoryTriggers_repositoryName - The name of the repository in which to test the triggers.

$sel:triggers:TestRepositoryTriggers', testRepositoryTriggers_triggers - The list of triggers to test.

Request Lenses

testRepositoryTriggers_repositoryName :: Lens' TestRepositoryTriggers Text

The name of the repository in which to test the triggers.

testRepositoryTriggers_triggers :: Lens' TestRepositoryTriggers [RepositoryTrigger]

The list of triggers to test.

Destructuring the Response

data TestRepositoryTriggersResponse

Represents the output of a test repository triggers operation.

See: newTestRepositoryTriggersResponse smart constructor.

Constructors

TestRepositoryTriggersResponse'

Fields

failedExecutions :: Maybe [RepositoryTriggerExecutionFailure]

The list of triggers that were not tested. This list provides the names of the triggers that could not be tested, separated by commas.

successfulExecutions :: Maybe [Text]

The list of triggers that were successfully tested. This list provides the names of the triggers that were successfully tested, separated by commas.

httpStatus :: Int

The response's http status code.

newTestRepositoryTriggersResponse

Arguments

:: Int
$sel:httpStatus:TestRepositoryTriggersResponse'

-> TestRepositoryTriggersResponse

Create a value of TestRepositoryTriggersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failedExecutions:TestRepositoryTriggersResponse', testRepositoryTriggersResponse_failedExecutions - The list of triggers that were not tested. This list provides the names of the triggers that could not be tested, separated by commas.

$sel:successfulExecutions:TestRepositoryTriggersResponse', testRepositoryTriggersResponse_successfulExecutions - The list of triggers that were successfully tested. This list provides the names of the triggers that were successfully tested, separated by commas.

$sel:httpStatus:TestRepositoryTriggersResponse', testRepositoryTriggersResponse_httpStatus - The response's http status code.

Response Lenses

testRepositoryTriggersResponse_failedExecutions :: Lens' TestRepositoryTriggersResponse (Maybe [RepositoryTriggerExecutionFailure])

The list of triggers that were not tested. This list provides the names of the triggers that could not be tested, separated by commas.

testRepositoryTriggersResponse_successfulExecutions :: Lens' TestRepositoryTriggersResponse (Maybe [Text])

The list of triggers that were successfully tested. This list provides the names of the triggers that were successfully tested, separated by commas.

testRepositoryTriggersResponse_httpStatus :: Lens' TestRepositoryTriggersResponse Int

The response's http status code.