Network.AWS.CodeCommit.Types.ApprovalRule

Description

Documentation

data ApprovalRule

Returns information about an approval rule.

See: newApprovalRule smart constructor.

Constructors

ApprovalRule'

Fields

lastModifiedDate :: Maybe POSIX

The date the approval rule was most recently changed, in timestamp format.

approvalRuleContent :: Maybe Text

The content of the approval rule.

creationDate :: Maybe POSIX

The date the approval rule was created, in timestamp format.

originApprovalRuleTemplate :: Maybe OriginApprovalRuleTemplate

The approval rule template used to create the rule.

ruleContentSha256 :: Maybe Text

The SHA-256 hash signature for the content of the approval rule.

approvalRuleId :: Maybe Text

The system-generated ID of the approval rule.

approvalRuleName :: Maybe Text

The name of the approval rule.

lastModifiedUser :: Maybe Text

The Amazon Resource Name (ARN) of the user who made the most recent changes to the approval rule.

newApprovalRule :: ApprovalRule

Create a value of ApprovalRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedDate:ApprovalRule', approvalRule_lastModifiedDate - The date the approval rule was most recently changed, in timestamp format.

$sel:approvalRuleContent:ApprovalRule', approvalRule_approvalRuleContent - The content of the approval rule.

$sel:creationDate:ApprovalRule', approvalRule_creationDate - The date the approval rule was created, in timestamp format.

$sel:originApprovalRuleTemplate:ApprovalRule', approvalRule_originApprovalRuleTemplate - The approval rule template used to create the rule.

$sel:ruleContentSha256:ApprovalRule', approvalRule_ruleContentSha256 - The SHA-256 hash signature for the content of the approval rule.

$sel:approvalRuleId:ApprovalRule', approvalRule_approvalRuleId - The system-generated ID of the approval rule.

$sel:approvalRuleName:ApprovalRule', approvalRule_approvalRuleName - The name of the approval rule.

$sel:lastModifiedUser:ApprovalRule', approvalRule_lastModifiedUser - The Amazon Resource Name (ARN) of the user who made the most recent changes to the approval rule.

approvalRule_lastModifiedDate :: Lens' ApprovalRule (Maybe UTCTime)

The date the approval rule was most recently changed, in timestamp format.

approvalRule_approvalRuleContent :: Lens' ApprovalRule (Maybe Text)

The content of the approval rule.

approvalRule_creationDate :: Lens' ApprovalRule (Maybe UTCTime)

The date the approval rule was created, in timestamp format.

approvalRule_originApprovalRuleTemplate :: Lens' ApprovalRule (Maybe OriginApprovalRuleTemplate)

The approval rule template used to create the rule.

approvalRule_ruleContentSha256 :: Lens' ApprovalRule (Maybe Text)

The SHA-256 hash signature for the content of the approval rule.

approvalRule_approvalRuleId :: Lens' ApprovalRule (Maybe Text)

The system-generated ID of the approval rule.

approvalRule_approvalRuleName :: Lens' ApprovalRule (Maybe Text)

The name of the approval rule.

approvalRule_lastModifiedUser :: Lens' ApprovalRule (Maybe Text)

The Amazon Resource Name (ARN) of the user who made the most recent changes to the approval rule.