Network.AWS.CodeCommit.Types.ApprovalRuleTemplate

Description

Documentation

data ApprovalRuleTemplate

Returns information about an approval rule template.

See: newApprovalRuleTemplate smart constructor.

Constructors

ApprovalRuleTemplate'

Fields

lastModifiedDate :: Maybe POSIX

The date the approval rule template was most recently changed, in timestamp format.

approvalRuleTemplateId :: Maybe Text

The system-generated ID of the approval rule template.

approvalRuleTemplateName :: Maybe Text

The name of the approval rule template.

creationDate :: Maybe POSIX

The date the approval rule template was created, in timestamp format.

approvalRuleTemplateDescription :: Maybe Text

The description of the approval rule template.

ruleContentSha256 :: Maybe Text

The SHA-256 hash signature for the content of the approval rule template.

approvalRuleTemplateContent :: Maybe Text

The content of the approval rule template.

lastModifiedUser :: Maybe Text

The Amazon Resource Name (ARN) of the user who made the most recent changes to the approval rule template.

newApprovalRuleTemplate :: ApprovalRuleTemplate

Create a value of ApprovalRuleTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedDate:ApprovalRuleTemplate', approvalRuleTemplate_lastModifiedDate - The date the approval rule template was most recently changed, in timestamp format.

$sel:approvalRuleTemplateId:ApprovalRuleTemplate', approvalRuleTemplate_approvalRuleTemplateId - The system-generated ID of the approval rule template.

$sel:approvalRuleTemplateName:ApprovalRuleTemplate', approvalRuleTemplate_approvalRuleTemplateName - The name of the approval rule template.

$sel:creationDate:ApprovalRuleTemplate', approvalRuleTemplate_creationDate - The date the approval rule template was created, in timestamp format.

$sel:approvalRuleTemplateDescription:ApprovalRuleTemplate', approvalRuleTemplate_approvalRuleTemplateDescription - The description of the approval rule template.

$sel:ruleContentSha256:ApprovalRuleTemplate', approvalRuleTemplate_ruleContentSha256 - The SHA-256 hash signature for the content of the approval rule template.

$sel:approvalRuleTemplateContent:ApprovalRuleTemplate', approvalRuleTemplate_approvalRuleTemplateContent - The content of the approval rule template.

$sel:lastModifiedUser:ApprovalRuleTemplate', approvalRuleTemplate_lastModifiedUser - The Amazon Resource Name (ARN) of the user who made the most recent changes to the approval rule template.

approvalRuleTemplate_lastModifiedDate :: Lens' ApprovalRuleTemplate (Maybe UTCTime)

The date the approval rule template was most recently changed, in timestamp format.

approvalRuleTemplate_approvalRuleTemplateId :: Lens' ApprovalRuleTemplate (Maybe Text)

The system-generated ID of the approval rule template.

approvalRuleTemplate_approvalRuleTemplateName :: Lens' ApprovalRuleTemplate (Maybe Text)

The name of the approval rule template.

approvalRuleTemplate_creationDate :: Lens' ApprovalRuleTemplate (Maybe UTCTime)

The date the approval rule template was created, in timestamp format.

approvalRuleTemplate_approvalRuleTemplateDescription :: Lens' ApprovalRuleTemplate (Maybe Text)

The description of the approval rule template.

approvalRuleTemplate_ruleContentSha256 :: Lens' ApprovalRuleTemplate (Maybe Text)

The SHA-256 hash signature for the content of the approval rule template.

approvalRuleTemplate_approvalRuleTemplateContent :: Lens' ApprovalRuleTemplate (Maybe Text)

The content of the approval rule template.

approvalRuleTemplate_lastModifiedUser :: Lens' ApprovalRuleTemplate (Maybe Text)

The Amazon Resource Name (ARN) of the user who made the most recent changes to the approval rule template.