Network.AWS.CodeCommit.Types.CommentsForPullRequest

Description

Documentation

data CommentsForPullRequest

Returns information about comments on a pull request.

See: newCommentsForPullRequest smart constructor.

Constructors

CommentsForPullRequest'

Fields

beforeBlobId :: Maybe Text

The full blob ID of the file on which you want to comment on the destination commit.

repositoryName :: Maybe Text

The name of the repository that contains the pull request.

beforeCommitId :: Maybe Text

The full commit ID of the commit that was the tip of the destination branch when the pull request was created. This commit is superceded by the after commit in the source branch when and if you merge the source branch into the destination branch.

afterBlobId :: Maybe Text

The full blob ID of the file on which you want to comment on the source commit.

pullRequestId :: Maybe Text

The system-generated ID of the pull request.

afterCommitId :: Maybe Text

The full commit ID of the commit that was the tip of the source branch at the time the comment was made.

comments :: Maybe [Comment]

An array of comment objects. Each comment object contains information about a comment on the pull request.

location :: Maybe Location

Location information about the comment on the pull request, including the file name, line number, and whether the version of the file where the comment was made is BEFORE (destination branch) or AFTER (source branch).

newCommentsForPullRequest :: CommentsForPullRequest

Create a value of CommentsForPullRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:beforeBlobId:CommentsForPullRequest', commentsForPullRequest_beforeBlobId - The full blob ID of the file on which you want to comment on the destination commit.

$sel:repositoryName:CommentsForPullRequest', commentsForPullRequest_repositoryName - The name of the repository that contains the pull request.

$sel:beforeCommitId:CommentsForPullRequest', commentsForPullRequest_beforeCommitId - The full commit ID of the commit that was the tip of the destination branch when the pull request was created. This commit is superceded by the after commit in the source branch when and if you merge the source branch into the destination branch.

$sel:afterBlobId:CommentsForPullRequest', commentsForPullRequest_afterBlobId - The full blob ID of the file on which you want to comment on the source commit.

$sel:pullRequestId:CommentsForPullRequest', commentsForPullRequest_pullRequestId - The system-generated ID of the pull request.

$sel:afterCommitId:CommentsForPullRequest', commentsForPullRequest_afterCommitId - The full commit ID of the commit that was the tip of the source branch at the time the comment was made.

$sel:comments:CommentsForPullRequest', commentsForPullRequest_comments - An array of comment objects. Each comment object contains information about a comment on the pull request.

$sel:location:CommentsForPullRequest', commentsForPullRequest_location - Location information about the comment on the pull request, including the file name, line number, and whether the version of the file where the comment was made is BEFORE (destination branch) or AFTER (source branch).

commentsForPullRequest_beforeBlobId :: Lens' CommentsForPullRequest (Maybe Text)

The full blob ID of the file on which you want to comment on the destination commit.

commentsForPullRequest_repositoryName :: Lens' CommentsForPullRequest (Maybe Text)

The name of the repository that contains the pull request.

commentsForPullRequest_beforeCommitId :: Lens' CommentsForPullRequest (Maybe Text)

The full commit ID of the commit that was the tip of the destination branch when the pull request was created. This commit is superceded by the after commit in the source branch when and if you merge the source branch into the destination branch.

commentsForPullRequest_afterBlobId :: Lens' CommentsForPullRequest (Maybe Text)

The full blob ID of the file on which you want to comment on the source commit.

commentsForPullRequest_pullRequestId :: Lens' CommentsForPullRequest (Maybe Text)

The system-generated ID of the pull request.

commentsForPullRequest_afterCommitId :: Lens' CommentsForPullRequest (Maybe Text)

The full commit ID of the commit that was the tip of the source branch at the time the comment was made.

commentsForPullRequest_comments :: Lens' CommentsForPullRequest (Maybe [Comment])

An array of comment objects. Each comment object contains information about a comment on the pull request.

commentsForPullRequest_location :: Lens' CommentsForPullRequest (Maybe Location)

Location information about the comment on the pull request, including the file name, line number, and whether the version of the file where the comment was made is BEFORE (destination branch) or AFTER (source branch).