Network.AWS.CodeCommit.Types.Commit

Description

Documentation

data Commit

Returns information about a specific commit.

See: newCommit smart constructor.

Constructors

Commit'

Fields

parents :: Maybe [Text]

A list of parent commits for the specified commit. Each parent commit ID is the full commit ID.

commitId :: Maybe Text

The full SHA ID of the specified commit.

additionalData :: Maybe Text

Any other data associated with the specified commit.

message :: Maybe Text

The commit message associated with the specified commit.

treeId :: Maybe Text

Tree information for the specified commit.

author :: Maybe UserInfo

Information about the author of the specified commit. Information includes the date in timestamp format with GMT offset, the name of the author, and the email address for the author, as configured in Git.

committer :: Maybe UserInfo

Information about the person who committed the specified commit, also known as the committer. Information includes the date in timestamp format with GMT offset, the name of the committer, and the email address for the committer, as configured in Git.

For more information about the difference between an author and a committer in Git, see Viewing the Commit History in Pro Git by Scott Chacon and Ben Straub.

newCommit :: Commit

Create a value of Commit with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parents:Commit', commit_parents - A list of parent commits for the specified commit. Each parent commit ID is the full commit ID.

$sel:commitId:Commit', commit_commitId - The full SHA ID of the specified commit.

$sel:additionalData:Commit', commit_additionalData - Any other data associated with the specified commit.

$sel:message:Commit', commit_message - The commit message associated with the specified commit.

$sel:treeId:Commit', commit_treeId - Tree information for the specified commit.

$sel:author:Commit', commit_author - Information about the author of the specified commit. Information includes the date in timestamp format with GMT offset, the name of the author, and the email address for the author, as configured in Git.

$sel:committer:Commit', commit_committer - Information about the person who committed the specified commit, also known as the committer. Information includes the date in timestamp format with GMT offset, the name of the committer, and the email address for the committer, as configured in Git.

For more information about the difference between an author and a committer in Git, see Viewing the Commit History in Pro Git by Scott Chacon and Ben Straub.

commit_parents :: Lens' Commit (Maybe [Text])

A list of parent commits for the specified commit. Each parent commit ID is the full commit ID.

commit_commitId :: Lens' Commit (Maybe Text)

The full SHA ID of the specified commit.

commit_additionalData :: Lens' Commit (Maybe Text)

Any other data associated with the specified commit.

commit_message :: Lens' Commit (Maybe Text)

The commit message associated with the specified commit.

commit_treeId :: Lens' Commit (Maybe Text)

Tree information for the specified commit.

commit_author :: Lens' Commit (Maybe UserInfo)

Information about the author of the specified commit. Information includes the date in timestamp format with GMT offset, the name of the author, and the email address for the author, as configured in Git.

commit_committer :: Lens' Commit (Maybe UserInfo)

Information about the person who committed the specified commit, also known as the committer. Information includes the date in timestamp format with GMT offset, the name of the committer, and the email address for the committer, as configured in Git.

For more information about the difference between an author and a committer in Git, see Viewing the Commit History in Pro Git by Scott Chacon and Ben Straub.