Network.AWS.CodeCommit.Types.ConflictMetadata

Description

Documentation

data ConflictMetadata

Information about the metadata for a conflict in a merge operation.

See: newConflictMetadata smart constructor.

Constructors

ConflictMetadata'

Fields

mergeOperations :: Maybe MergeOperations

Whether an add, modify, or delete operation caused the conflict between the source and destination of the merge.

filePath :: Maybe Text

The path of the file that contains conflicts.

isBinaryFile :: Maybe IsBinaryFile

A boolean value (true or false) indicating whether the file is binary or textual in the source, destination, and base of the merge.

objectTypeConflict :: Maybe Bool

A boolean value (true or false) indicating whether there are conflicts between the branches in the object type of a file, folder, or submodule.

fileModeConflict :: Maybe Bool

A boolean value indicating whether there are conflicts in the file mode of a file.

numberOfConflicts :: Maybe Int

The number of conflicts, including both hunk conflicts and metadata conflicts.

contentConflict :: Maybe Bool

A boolean value indicating whether there are conflicts in the content of a file.

fileModes :: Maybe FileModes

The file modes of the file in the source, destination, and base of the merge.

objectTypes :: Maybe ObjectTypes

Information about any object type conflicts in a merge operation.

fileSizes :: Maybe FileSizes

The file sizes of the file in the source, destination, and base of the merge.

newConflictMetadata :: ConflictMetadata

Create a value of ConflictMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:mergeOperations:ConflictMetadata', conflictMetadata_mergeOperations - Whether an add, modify, or delete operation caused the conflict between the source and destination of the merge.

$sel:filePath:ConflictMetadata', conflictMetadata_filePath - The path of the file that contains conflicts.

$sel:isBinaryFile:ConflictMetadata', conflictMetadata_isBinaryFile - A boolean value (true or false) indicating whether the file is binary or textual in the source, destination, and base of the merge.

$sel:objectTypeConflict:ConflictMetadata', conflictMetadata_objectTypeConflict - A boolean value (true or false) indicating whether there are conflicts between the branches in the object type of a file, folder, or submodule.

$sel:fileModeConflict:ConflictMetadata', conflictMetadata_fileModeConflict - A boolean value indicating whether there are conflicts in the file mode of a file.

$sel:numberOfConflicts:ConflictMetadata', conflictMetadata_numberOfConflicts - The number of conflicts, including both hunk conflicts and metadata conflicts.

$sel:contentConflict:ConflictMetadata', conflictMetadata_contentConflict - A boolean value indicating whether there are conflicts in the content of a file.

$sel:fileModes:ConflictMetadata', conflictMetadata_fileModes - The file modes of the file in the source, destination, and base of the merge.

$sel:objectTypes:ConflictMetadata', conflictMetadata_objectTypes - Information about any object type conflicts in a merge operation.

$sel:fileSizes:ConflictMetadata', conflictMetadata_fileSizes - The file sizes of the file in the source, destination, and base of the merge.

conflictMetadata_mergeOperations :: Lens' ConflictMetadata (Maybe MergeOperations)

Whether an add, modify, or delete operation caused the conflict between the source and destination of the merge.

conflictMetadata_filePath :: Lens' ConflictMetadata (Maybe Text)

The path of the file that contains conflicts.

conflictMetadata_isBinaryFile :: Lens' ConflictMetadata (Maybe IsBinaryFile)

A boolean value (true or false) indicating whether the file is binary or textual in the source, destination, and base of the merge.

conflictMetadata_objectTypeConflict :: Lens' ConflictMetadata (Maybe Bool)

A boolean value (true or false) indicating whether there are conflicts between the branches in the object type of a file, folder, or submodule.

conflictMetadata_fileModeConflict :: Lens' ConflictMetadata (Maybe Bool)

A boolean value indicating whether there are conflicts in the file mode of a file.

conflictMetadata_numberOfConflicts :: Lens' ConflictMetadata (Maybe Int)

The number of conflicts, including both hunk conflicts and metadata conflicts.

conflictMetadata_contentConflict :: Lens' ConflictMetadata (Maybe Bool)

A boolean value indicating whether there are conflicts in the content of a file.

conflictMetadata_fileModes :: Lens' ConflictMetadata (Maybe FileModes)

The file modes of the file in the source, destination, and base of the merge.

conflictMetadata_objectTypes :: Lens' ConflictMetadata (Maybe ObjectTypes)

Information about any object type conflicts in a merge operation.

conflictMetadata_fileSizes :: Lens' ConflictMetadata (Maybe FileSizes)

The file sizes of the file in the source, destination, and base of the merge.