Network.AWS.CodeCommit.Types.Evaluation

Description

Documentation

data Evaluation

Returns information about the approval rules applied to a pull request and whether conditions have been met.

See: newEvaluation smart constructor.

Constructors

Evaluation'

Fields

approvalRulesSatisfied :: Maybe [Text]

The names of the approval rules that have had their conditions met.

approvalRulesNotSatisfied :: Maybe [Text]

The names of the approval rules that have not had their conditions met.

approved :: Maybe Bool

Whether the state of the pull request is approved.

overridden :: Maybe Bool

Whether the approval rule requirements for the pull request have been overridden and no longer need to be met.

newEvaluation :: Evaluation

Create a value of Evaluation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:approvalRulesSatisfied:Evaluation', evaluation_approvalRulesSatisfied - The names of the approval rules that have had their conditions met.

$sel:approvalRulesNotSatisfied:Evaluation', evaluation_approvalRulesNotSatisfied - The names of the approval rules that have not had their conditions met.

$sel:approved:Evaluation', evaluation_approved - Whether the state of the pull request is approved.

$sel:overridden:Evaluation', evaluation_overridden - Whether the approval rule requirements for the pull request have been overridden and no longer need to be met.

evaluation_approvalRulesSatisfied :: Lens' Evaluation (Maybe [Text])

The names of the approval rules that have had their conditions met.

evaluation_approvalRulesNotSatisfied :: Lens' Evaluation (Maybe [Text])

The names of the approval rules that have not had their conditions met.

evaluation_approved :: Lens' Evaluation (Maybe Bool)

Whether the state of the pull request is approved.

evaluation_overridden :: Lens' Evaluation (Maybe Bool)

Whether the approval rule requirements for the pull request have been overridden and no longer need to be met.