Network.AWS.CodeCommit.Types.FileModes

Description

Documentation

data FileModes

Information about file modes in a merge or pull request.

See: newFileModes smart constructor.

Constructors

FileModes'

Fields

destination :: Maybe FileModeTypeEnum

The file mode of a file in the destination of a merge or pull request.

base :: Maybe FileModeTypeEnum

The file mode of a file in the base of a merge or pull request.

source :: Maybe FileModeTypeEnum

The file mode of a file in the source of a merge or pull request.

newFileModes :: FileModes

Create a value of FileModes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:destination:FileModes', fileModes_destination - The file mode of a file in the destination of a merge or pull request.

$sel:base:FileModes', fileModes_base - The file mode of a file in the base of a merge or pull request.

$sel:source:FileModes', fileModes_source - The file mode of a file in the source of a merge or pull request.

fileModes_destination :: Lens' FileModes (Maybe FileModeTypeEnum)

The file mode of a file in the destination of a merge or pull request.

fileModes_base :: Lens' FileModes (Maybe FileModeTypeEnum)

The file mode of a file in the base of a merge or pull request.

fileModes_source :: Lens' FileModes (Maybe FileModeTypeEnum)

The file mode of a file in the source of a merge or pull request.