Network.AWS.CodeCommit.Types.FileSizes

Description

Documentation

data FileSizes

Information about the size of files in a merge or pull request.

See: newFileSizes smart constructor.

Constructors

FileSizes'

Fields

source :: Maybe Integer

The size of a file in the source of a merge or pull request.

destination :: Maybe Integer

The size of a file in the destination of a merge or pull request.

base :: Maybe Integer

The size of a file in the base of a merge or pull request.

newFileSizes :: FileSizes

Create a value of FileSizes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:source:FileSizes', fileSizes_source - The size of a file in the source of a merge or pull request.

$sel:destination:FileSizes', fileSizes_destination - The size of a file in the destination of a merge or pull request.

$sel:base:FileSizes', fileSizes_base - The size of a file in the base of a merge or pull request.

fileSizes_source :: Lens' FileSizes (Maybe Integer)

The size of a file in the source of a merge or pull request.

fileSizes_destination :: Lens' FileSizes (Maybe Integer)

The size of a file in the destination of a merge or pull request.

fileSizes_base :: Lens' FileSizes (Maybe Integer)

The size of a file in the base of a merge or pull request.