Network.AWS.CodeCommit.Types.MergeHunkDetail

Description

Documentation

data MergeHunkDetail

Information about the details of a merge hunk that contains a conflict in a merge or pull request operation.

See: newMergeHunkDetail smart constructor.

Constructors

MergeHunkDetail'

Fields

startLine :: Maybe Int

The start position of the hunk in the merge result.

endLine :: Maybe Int

The end position of the hunk in the merge result.

hunkContent :: Maybe Text

The base-64 encoded content of the hunk merged region that might contain a conflict.

newMergeHunkDetail :: MergeHunkDetail

Create a value of MergeHunkDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:startLine:MergeHunkDetail', mergeHunkDetail_startLine - The start position of the hunk in the merge result.

$sel:endLine:MergeHunkDetail', mergeHunkDetail_endLine - The end position of the hunk in the merge result.

$sel:hunkContent:MergeHunkDetail', mergeHunkDetail_hunkContent - The base-64 encoded content of the hunk merged region that might contain a conflict.

mergeHunkDetail_startLine :: Lens' MergeHunkDetail (Maybe Int)

The start position of the hunk in the merge result.

mergeHunkDetail_endLine :: Lens' MergeHunkDetail (Maybe Int)

The end position of the hunk in the merge result.

mergeHunkDetail_hunkContent :: Lens' MergeHunkDetail (Maybe Text)

The base-64 encoded content of the hunk merged region that might contain a conflict.