Network.AWS.CodeCommit.Types.PullRequestEvent

Description

Documentation

data PullRequestEvent

Returns information about a pull request event.

See: newPullRequestEvent smart constructor.

Constructors

PullRequestEvent'

Fields

pullRequestMergedStateChangedEventMetadata :: Maybe PullRequestMergedStateChangedEventMetadata

Information about the change in mergability state for the pull request event.

pullRequestSourceReferenceUpdatedEventMetadata :: Maybe PullRequestSourceReferenceUpdatedEventMetadata

Information about the updated source branch for the pull request event.

approvalStateChangedEventMetadata :: Maybe ApprovalStateChangedEventMetadata

Information about an approval state change for a pull request.

eventDate :: Maybe POSIX

The day and time of the pull request event, in timestamp format.

pullRequestEventType :: Maybe PullRequestEventType

The type of the pull request event (for example, a status change event (PULL_REQUEST_STATUS_CHANGED) or update event (PULL_REQUEST_SOURCE_REFERENCE_UPDATED)).

pullRequestCreatedEventMetadata :: Maybe PullRequestCreatedEventMetadata

Information about the source and destination branches for the pull request.

pullRequestId :: Maybe Text

The system-generated ID of the pull request.

actorArn :: Maybe Text

The Amazon Resource Name (ARN) of the user whose actions resulted in the event. Examples include updating the pull request with more commits or changing the status of a pull request.

approvalRuleOverriddenEventMetadata :: Maybe ApprovalRuleOverriddenEventMetadata

Information about an approval rule override event for a pull request.

pullRequestStatusChangedEventMetadata :: Maybe PullRequestStatusChangedEventMetadata

Information about the change in status for the pull request event.

approvalRuleEventMetadata :: Maybe ApprovalRuleEventMetadata

Information about a pull request event.

newPullRequestEvent :: PullRequestEvent

Create a value of PullRequestEvent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pullRequestMergedStateChangedEventMetadata:PullRequestEvent', pullRequestEvent_pullRequestMergedStateChangedEventMetadata - Information about the change in mergability state for the pull request event.

$sel:pullRequestSourceReferenceUpdatedEventMetadata:PullRequestEvent', pullRequestEvent_pullRequestSourceReferenceUpdatedEventMetadata - Information about the updated source branch for the pull request event.

$sel:approvalStateChangedEventMetadata:PullRequestEvent', pullRequestEvent_approvalStateChangedEventMetadata - Information about an approval state change for a pull request.

$sel:eventDate:PullRequestEvent', pullRequestEvent_eventDate - The day and time of the pull request event, in timestamp format.

$sel:pullRequestEventType:PullRequestEvent', pullRequestEvent_pullRequestEventType - The type of the pull request event (for example, a status change event (PULL_REQUEST_STATUS_CHANGED) or update event (PULL_REQUEST_SOURCE_REFERENCE_UPDATED)).

$sel:pullRequestCreatedEventMetadata:PullRequestEvent', pullRequestEvent_pullRequestCreatedEventMetadata - Information about the source and destination branches for the pull request.

$sel:pullRequestId:PullRequestEvent', pullRequestEvent_pullRequestId - The system-generated ID of the pull request.

$sel:actorArn:PullRequestEvent', pullRequestEvent_actorArn - The Amazon Resource Name (ARN) of the user whose actions resulted in the event. Examples include updating the pull request with more commits or changing the status of a pull request.

$sel:approvalRuleOverriddenEventMetadata:PullRequestEvent', pullRequestEvent_approvalRuleOverriddenEventMetadata - Information about an approval rule override event for a pull request.

$sel:pullRequestStatusChangedEventMetadata:PullRequestEvent', pullRequestEvent_pullRequestStatusChangedEventMetadata - Information about the change in status for the pull request event.

$sel:approvalRuleEventMetadata:PullRequestEvent', pullRequestEvent_approvalRuleEventMetadata - Information about a pull request event.

pullRequestEvent_pullRequestMergedStateChangedEventMetadata :: Lens' PullRequestEvent (Maybe PullRequestMergedStateChangedEventMetadata)

Information about the change in mergability state for the pull request event.

pullRequestEvent_pullRequestSourceReferenceUpdatedEventMetadata :: Lens' PullRequestEvent (Maybe PullRequestSourceReferenceUpdatedEventMetadata)

Information about the updated source branch for the pull request event.

pullRequestEvent_approvalStateChangedEventMetadata :: Lens' PullRequestEvent (Maybe ApprovalStateChangedEventMetadata)

Information about an approval state change for a pull request.

pullRequestEvent_eventDate :: Lens' PullRequestEvent (Maybe UTCTime)

The day and time of the pull request event, in timestamp format.

pullRequestEvent_pullRequestEventType :: Lens' PullRequestEvent (Maybe PullRequestEventType)

The type of the pull request event (for example, a status change event (PULL_REQUEST_STATUS_CHANGED) or update event (PULL_REQUEST_SOURCE_REFERENCE_UPDATED)).

pullRequestEvent_pullRequestCreatedEventMetadata :: Lens' PullRequestEvent (Maybe PullRequestCreatedEventMetadata)

Information about the source and destination branches for the pull request.

pullRequestEvent_pullRequestId :: Lens' PullRequestEvent (Maybe Text)

The system-generated ID of the pull request.

pullRequestEvent_actorArn :: Lens' PullRequestEvent (Maybe Text)

The Amazon Resource Name (ARN) of the user whose actions resulted in the event. Examples include updating the pull request with more commits or changing the status of a pull request.

pullRequestEvent_approvalRuleOverriddenEventMetadata :: Lens' PullRequestEvent (Maybe ApprovalRuleOverriddenEventMetadata)

Information about an approval rule override event for a pull request.

pullRequestEvent_pullRequestStatusChangedEventMetadata :: Lens' PullRequestEvent (Maybe PullRequestStatusChangedEventMetadata)

Information about the change in status for the pull request event.

pullRequestEvent_approvalRuleEventMetadata :: Lens' PullRequestEvent (Maybe ApprovalRuleEventMetadata)

Information about a pull request event.