Network.AWS.CodeCommit.Types.PullRequestTarget

Description

Documentation

data PullRequestTarget

Returns information about a pull request target.

See: newPullRequestTarget smart constructor.

Constructors

PullRequestTarget'

Fields

destinationReference :: Maybe Text

The branch of the repository where the pull request changes are merged. Also known as the destination branch.

sourceCommit :: Maybe Text

The full commit ID of the tip of the source branch used to create the pull request. If the pull request branch is updated by a push while the pull request is open, the commit ID changes to reflect the new tip of the branch.

mergeMetadata :: Maybe MergeMetadata

Returns metadata about the state of the merge, including whether the merge has been made.

repositoryName :: Maybe Text

The name of the repository that contains the pull request source and destination branches.

sourceReference :: Maybe Text

The branch of the repository that contains the changes for the pull request. Also known as the source branch.

mergeBase :: Maybe Text

The commit ID of the most recent commit that the source branch and the destination branch have in common.

destinationCommit :: Maybe Text

The full commit ID that is the tip of the destination branch. This is the commit where the pull request was or will be merged.

newPullRequestTarget :: PullRequestTarget

Create a value of PullRequestTarget with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:destinationReference:PullRequestTarget', pullRequestTarget_destinationReference - The branch of the repository where the pull request changes are merged. Also known as the destination branch.

$sel:sourceCommit:PullRequestTarget', pullRequestTarget_sourceCommit - The full commit ID of the tip of the source branch used to create the pull request. If the pull request branch is updated by a push while the pull request is open, the commit ID changes to reflect the new tip of the branch.

$sel:mergeMetadata:PullRequestTarget', pullRequestTarget_mergeMetadata - Returns metadata about the state of the merge, including whether the merge has been made.

$sel:repositoryName:PullRequestTarget', pullRequestTarget_repositoryName - The name of the repository that contains the pull request source and destination branches.

$sel:sourceReference:PullRequestTarget', pullRequestTarget_sourceReference - The branch of the repository that contains the changes for the pull request. Also known as the source branch.

$sel:mergeBase:PullRequestTarget', pullRequestTarget_mergeBase - The commit ID of the most recent commit that the source branch and the destination branch have in common.

$sel:destinationCommit:PullRequestTarget', pullRequestTarget_destinationCommit - The full commit ID that is the tip of the destination branch. This is the commit where the pull request was or will be merged.

pullRequestTarget_destinationReference :: Lens' PullRequestTarget (Maybe Text)

The branch of the repository where the pull request changes are merged. Also known as the destination branch.

pullRequestTarget_sourceCommit :: Lens' PullRequestTarget (Maybe Text)

The full commit ID of the tip of the source branch used to create the pull request. If the pull request branch is updated by a push while the pull request is open, the commit ID changes to reflect the new tip of the branch.

pullRequestTarget_mergeMetadata :: Lens' PullRequestTarget (Maybe MergeMetadata)

Returns metadata about the state of the merge, including whether the merge has been made.

pullRequestTarget_repositoryName :: Lens' PullRequestTarget (Maybe Text)

The name of the repository that contains the pull request source and destination branches.

pullRequestTarget_sourceReference :: Lens' PullRequestTarget (Maybe Text)

The branch of the repository that contains the changes for the pull request. Also known as the source branch.

pullRequestTarget_mergeBase :: Lens' PullRequestTarget (Maybe Text)

The commit ID of the most recent commit that the source branch and the destination branch have in common.

pullRequestTarget_destinationCommit :: Lens' PullRequestTarget (Maybe Text)

The full commit ID that is the tip of the destination branch. This is the commit where the pull request was or will be merged.