Network.AWS.CodeCommit.Types.ReactionValueFormats

Description

Documentation

data ReactionValueFormats

Information about the values for reactions to a comment. AWS CodeCommit supports a limited set of reactions.

See: newReactionValueFormats smart constructor.

Constructors

ReactionValueFormats'

Fields

shortCode :: Maybe Text

The emoji short code for the reaction. Short codes are interpreted slightly differently on different operating systems.

unicode :: Maybe Text

The Unicode codepoint for the reaction.

emoji :: Maybe Text

The Emoji Version 1.0 graphic of the reaction. These graphics are interpreted slightly differently on different operating systems.

newReactionValueFormats :: ReactionValueFormats

Create a value of ReactionValueFormats with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:shortCode:ReactionValueFormats', reactionValueFormats_shortCode - The emoji short code for the reaction. Short codes are interpreted slightly differently on different operating systems.

$sel:unicode:ReactionValueFormats', reactionValueFormats_unicode - The Unicode codepoint for the reaction.

$sel:emoji:ReactionValueFormats', reactionValueFormats_emoji - The Emoji Version 1.0 graphic of the reaction. These graphics are interpreted slightly differently on different operating systems.

reactionValueFormats_shortCode :: Lens' ReactionValueFormats (Maybe Text)

The emoji short code for the reaction. Short codes are interpreted slightly differently on different operating systems.

reactionValueFormats_unicode :: Lens' ReactionValueFormats (Maybe Text)

The Unicode codepoint for the reaction.

reactionValueFormats_emoji :: Lens' ReactionValueFormats (Maybe Text)

The Emoji Version 1.0 graphic of the reaction. These graphics are interpreted slightly differently on different operating systems.