Network.AWS.CodeCommit.Types.ReplaceContentEntry

Description

Documentation

data ReplaceContentEntry

Information about a replacement content entry in the conflict of a merge or pull request operation.

See: newReplaceContentEntry smart constructor.

Constructors

ReplaceContentEntry'

Fields

content :: Maybe Base64

The base-64 encoded content to use when the replacement type is USE_NEW_CONTENT.

fileMode :: Maybe FileModeTypeEnum

The file mode to apply during conflict resoltion.

filePath :: Text

The path of the conflicting file.

replacementType :: ReplacementTypeEnum

The replacement type to use when determining how to resolve the conflict.

newReplaceContentEntry

Arguments

:: Text
$sel:filePath:ReplaceContentEntry'

-> ReplacementTypeEnum
$sel:replacementType:ReplaceContentEntry'

-> ReplaceContentEntry

Create a value of ReplaceContentEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:content:ReplaceContentEntry', replaceContentEntry_content - The base-64 encoded content to use when the replacement type is USE_NEW_CONTENT.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:fileMode:ReplaceContentEntry', replaceContentEntry_fileMode - The file mode to apply during conflict resoltion.

$sel:filePath:ReplaceContentEntry', replaceContentEntry_filePath - The path of the conflicting file.

$sel:replacementType:ReplaceContentEntry', replaceContentEntry_replacementType - The replacement type to use when determining how to resolve the conflict.

replaceContentEntry_content :: Lens' ReplaceContentEntry (Maybe ByteString)

The base-64 encoded content to use when the replacement type is USE_NEW_CONTENT.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

replaceContentEntry_fileMode :: Lens' ReplaceContentEntry (Maybe FileModeTypeEnum)

The file mode to apply during conflict resoltion.

replaceContentEntry_filePath :: Lens' ReplaceContentEntry Text

The path of the conflicting file.

replaceContentEntry_replacementType :: Lens' ReplaceContentEntry ReplacementTypeEnum

The replacement type to use when determining how to resolve the conflict.