Network.AWS.CodeCommit.Types.RepositoryMetadata

Description

Documentation

data RepositoryMetadata

Information about a repository.

See: newRepositoryMetadata smart constructor.

Constructors

RepositoryMetadata'

Fields

lastModifiedDate :: Maybe POSIX

The date and time the repository was last modified, in timestamp format.

defaultBranch :: Maybe Text

The repository's default branch name.

accountId :: Maybe Text

The ID of the AWS account associated with the repository.

cloneUrlHttp :: Maybe Text

The URL to use for cloning the repository over HTTPS.

cloneUrlSsh :: Maybe Text

The URL to use for cloning the repository over SSH.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the repository.

creationDate :: Maybe POSIX

The date and time the repository was created, in timestamp format.

repositoryName :: Maybe Text

The repository's name.

repositoryId :: Maybe Text

The ID of the repository.

repositoryDescription :: Maybe Text

A comment or description about the repository.

newRepositoryMetadata :: RepositoryMetadata

Create a value of RepositoryMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedDate:RepositoryMetadata', repositoryMetadata_lastModifiedDate - The date and time the repository was last modified, in timestamp format.

$sel:defaultBranch:RepositoryMetadata', repositoryMetadata_defaultBranch - The repository's default branch name.

$sel:accountId:RepositoryMetadata', repositoryMetadata_accountId - The ID of the AWS account associated with the repository.

$sel:cloneUrlHttp:RepositoryMetadata', repositoryMetadata_cloneUrlHttp - The URL to use for cloning the repository over HTTPS.

$sel:cloneUrlSsh:RepositoryMetadata', repositoryMetadata_cloneUrlSsh - The URL to use for cloning the repository over SSH.

$sel:arn:RepositoryMetadata', repositoryMetadata_arn - The Amazon Resource Name (ARN) of the repository.

$sel:creationDate:RepositoryMetadata', repositoryMetadata_creationDate - The date and time the repository was created, in timestamp format.

$sel:repositoryName:RepositoryMetadata', repositoryMetadata_repositoryName - The repository's name.

$sel:repositoryId:RepositoryMetadata', repositoryMetadata_repositoryId - The ID of the repository.

$sel:repositoryDescription:RepositoryMetadata', repositoryMetadata_repositoryDescription - A comment or description about the repository.

repositoryMetadata_lastModifiedDate :: Lens' RepositoryMetadata (Maybe UTCTime)

The date and time the repository was last modified, in timestamp format.

repositoryMetadata_defaultBranch :: Lens' RepositoryMetadata (Maybe Text)

The repository's default branch name.

repositoryMetadata_accountId :: Lens' RepositoryMetadata (Maybe Text)

The ID of the AWS account associated with the repository.

repositoryMetadata_cloneUrlHttp :: Lens' RepositoryMetadata (Maybe Text)

The URL to use for cloning the repository over HTTPS.

repositoryMetadata_cloneUrlSsh :: Lens' RepositoryMetadata (Maybe Text)

The URL to use for cloning the repository over SSH.

repositoryMetadata_arn :: Lens' RepositoryMetadata (Maybe Text)

The Amazon Resource Name (ARN) of the repository.

repositoryMetadata_creationDate :: Lens' RepositoryMetadata (Maybe UTCTime)

The date and time the repository was created, in timestamp format.

repositoryMetadata_repositoryName :: Lens' RepositoryMetadata (Maybe Text)

The repository's name.

repositoryMetadata_repositoryId :: Lens' RepositoryMetadata (Maybe Text)

The ID of the repository.

repositoryMetadata_repositoryDescription :: Lens' RepositoryMetadata (Maybe Text)

A comment or description about the repository.