Network.AWS.CodeCommit.Types.RepositoryTrigger

Description

Documentation

data RepositoryTrigger

Information about a trigger for a repository.

See: newRepositoryTrigger smart constructor.

Constructors

RepositoryTrigger'

Fields

customData :: Maybe Text

Any custom data associated with the trigger to be included in the information sent to the target of the trigger.

branches :: Maybe [Text]

The branches to be included in the trigger configuration. If you specify an empty array, the trigger applies to all branches.

Although no content is required in the array, you must include the array itself.

name :: Text

The name of the trigger.

destinationArn :: Text

The ARN of the resource that is the target for a trigger (for example, the ARN of a topic in Amazon SNS).

events :: [RepositoryTriggerEventEnum]

The repository events that cause the trigger to run actions in another service, such as sending a notification through Amazon SNS.

The valid value "all" cannot be used with any other values.

newRepositoryTrigger

Arguments

:: Text
$sel:name:RepositoryTrigger'

-> Text
$sel:destinationArn:RepositoryTrigger'

-> RepositoryTrigger

Create a value of RepositoryTrigger with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:customData:RepositoryTrigger', repositoryTrigger_customData - Any custom data associated with the trigger to be included in the information sent to the target of the trigger.

$sel:branches:RepositoryTrigger', repositoryTrigger_branches - The branches to be included in the trigger configuration. If you specify an empty array, the trigger applies to all branches.

Although no content is required in the array, you must include the array itself.

$sel:name:RepositoryTrigger', repositoryTrigger_name - The name of the trigger.

$sel:destinationArn:RepositoryTrigger', repositoryTrigger_destinationArn - The ARN of the resource that is the target for a trigger (for example, the ARN of a topic in Amazon SNS).

$sel:events:RepositoryTrigger', repositoryTrigger_events - The repository events that cause the trigger to run actions in another service, such as sending a notification through Amazon SNS.

The valid value "all" cannot be used with any other values.

repositoryTrigger_customData :: Lens' RepositoryTrigger (Maybe Text)

Any custom data associated with the trigger to be included in the information sent to the target of the trigger.

repositoryTrigger_branches :: Lens' RepositoryTrigger (Maybe [Text])

The branches to be included in the trigger configuration. If you specify an empty array, the trigger applies to all branches.

Although no content is required in the array, you must include the array itself.

repositoryTrigger_name :: Lens' RepositoryTrigger Text

The name of the trigger.

repositoryTrigger_destinationArn :: Lens' RepositoryTrigger Text

The ARN of the resource that is the target for a trigger (for example, the ARN of a topic in Amazon SNS).

repositoryTrigger_events :: Lens' RepositoryTrigger [RepositoryTriggerEventEnum]

The repository events that cause the trigger to run actions in another service, such as sending a notification through Amazon SNS.

The valid value "all" cannot be used with any other values.