Network.AWS.CodeCommit.Types.SymbolicLink

Description

Documentation

data SymbolicLink

Returns information about a symbolic link in a repository folder.

See: newSymbolicLink smart constructor.

Constructors

SymbolicLink'

Fields

absolutePath :: Maybe Text

The fully qualified path to the folder that contains the symbolic link.

fileMode :: Maybe FileModeTypeEnum

The file mode permissions of the blob that cotains information about the symbolic link.

blobId :: Maybe Text

The blob ID that contains the information about the symbolic link.

relativePath :: Maybe Text

The relative path of the symbolic link from the folder where the query originated.

newSymbolicLink :: SymbolicLink

Create a value of SymbolicLink with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:absolutePath:SymbolicLink', symbolicLink_absolutePath - The fully qualified path to the folder that contains the symbolic link.

$sel:fileMode:SymbolicLink', symbolicLink_fileMode - The file mode permissions of the blob that cotains information about the symbolic link.

$sel:blobId:SymbolicLink', symbolicLink_blobId - The blob ID that contains the information about the symbolic link.

$sel:relativePath:SymbolicLink', symbolicLink_relativePath - The relative path of the symbolic link from the folder where the query originated.

symbolicLink_absolutePath :: Lens' SymbolicLink (Maybe Text)

The fully qualified path to the folder that contains the symbolic link.

symbolicLink_fileMode :: Lens' SymbolicLink (Maybe FileModeTypeEnum)

The file mode permissions of the blob that cotains information about the symbolic link.

symbolicLink_blobId :: Lens' SymbolicLink (Maybe Text)

The blob ID that contains the information about the symbolic link.

symbolicLink_relativePath :: Lens' SymbolicLink (Maybe Text)

The relative path of the symbolic link from the folder where the query originated.