Network.AWS.CodeCommit.UpdatePullRequestApprovalState

Description

Updates the state of a user's approval on a pull request. The user is derived from the signed-in account when the request is made.

Creating a Request

data UpdatePullRequestApprovalState

See: newUpdatePullRequestApprovalState smart constructor.

Constructors

UpdatePullRequestApprovalState'

Fields

pullRequestId :: Text

The system-generated ID of the pull request.

revisionId :: Text

The system-generated ID of the revision.

approvalState :: ApprovalState

The approval state to associate with the user on the pull request.

newUpdatePullRequestApprovalState

Arguments

:: Text
$sel:pullRequestId:UpdatePullRequestApprovalState'

-> Text
$sel:revisionId:UpdatePullRequestApprovalState'

-> ApprovalState
$sel:approvalState:UpdatePullRequestApprovalState'

-> UpdatePullRequestApprovalState

Create a value of UpdatePullRequestApprovalState with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pullRequestId:UpdatePullRequestApprovalState', updatePullRequestApprovalState_pullRequestId - The system-generated ID of the pull request.

$sel:revisionId:UpdatePullRequestApprovalState', updatePullRequestApprovalState_revisionId - The system-generated ID of the revision.

$sel:approvalState:UpdatePullRequestApprovalState', updatePullRequestApprovalState_approvalState - The approval state to associate with the user on the pull request.

Request Lenses

updatePullRequestApprovalState_pullRequestId :: Lens' UpdatePullRequestApprovalState Text

The system-generated ID of the pull request.

updatePullRequestApprovalState_revisionId :: Lens' UpdatePullRequestApprovalState Text

The system-generated ID of the revision.

updatePullRequestApprovalState_approvalState :: Lens' UpdatePullRequestApprovalState ApprovalState

The approval state to associate with the user on the pull request.

Destructuring the Response

data UpdatePullRequestApprovalStateResponse

See: newUpdatePullRequestApprovalStateResponse smart constructor.

Constructors

UpdatePullRequestApprovalStateResponse'

newUpdatePullRequestApprovalStateResponse :: UpdatePullRequestApprovalStateResponse

Create a value of UpdatePullRequestApprovalStateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.