Network.AWS.CodeCommit.UpdateRepositoryName

Description

Renames a repository. The repository name must be unique across the calling AWS account. Repository names are limited to 100 alphanumeric, dash, and underscore characters, and cannot include certain characters. The suffix .git is prohibited. For more information about the limits on repository names, see Limits in the AWS CodeCommit User Guide.

Creating a Request

data UpdateRepositoryName

Represents the input of an update repository description operation.

See: newUpdateRepositoryName smart constructor.

Constructors

UpdateRepositoryName'

Fields

oldName :: Text

The current name of the repository.

newName' :: Text

The new name for the repository.

newUpdateRepositoryName

Arguments

:: Text
$sel:oldName:UpdateRepositoryName'

-> Text
$sel:newName':UpdateRepositoryName'

-> UpdateRepositoryName

Create a value of UpdateRepositoryName with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:oldName:UpdateRepositoryName', updateRepositoryName_oldName - The current name of the repository.

$sel:newName':UpdateRepositoryName', updateRepositoryName_newName - The new name for the repository.

Request Lenses

updateRepositoryName_oldName :: Lens' UpdateRepositoryName Text

The current name of the repository.

updateRepositoryName_newName :: Lens' UpdateRepositoryName Text

The new name for the repository.

Destructuring the Response

data UpdateRepositoryNameResponse

See: newUpdateRepositoryNameResponse smart constructor.

Constructors

UpdateRepositoryNameResponse'

newUpdateRepositoryNameResponse :: UpdateRepositoryNameResponse

Create a value of UpdateRepositoryNameResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.