Network.AWS.CodeDeploy.CreateDeploymentGroup

Description

Creates a deployment group to which application revisions are deployed.

Creating a Request

data CreateDeploymentGroup

Represents the input of a CreateDeploymentGroup operation.

See: newCreateDeploymentGroup smart constructor.

Constructors

CreateDeploymentGroup'

Fields

onPremisesTagSet :: Maybe OnPremisesTagSet

Information about groups of tags applied to on-premises instances. The deployment group includes only on-premises instances identified by all of the tag groups. Cannot be used in the same call as onPremisesInstanceTagFilters.

outdatedInstancesStrategy :: Maybe OutdatedInstancesStrategy

Indicates what happens when new EC2 instances are launched mid-deployment and do not receive the deployed application revision.

If this option is set to UPDATE or is unspecified, CodeDeploy initiates one or more 'auto-update outdated instances' deployments to apply the deployed application revision to the new EC2 instances.

If this option is set to IGNORE, CodeDeploy does not initiate a deployment to update the new EC2 instances. This may result in instances having different revisions.

deploymentConfigName :: Maybe Text

If specified, the deployment configuration name can be either one of the predefined configurations provided with AWS CodeDeploy or a custom deployment configuration that you create by calling the create deployment configuration operation.

CodeDeployDefault.OneAtATime is the default deployment configuration. It is used if a configuration isn't specified for the deployment or deployment group.

For more information about the predefined deployment configurations in AWS CodeDeploy, see Working with Deployment Configurations in CodeDeploy in the AWS CodeDeploy User Guide.

autoRollbackConfiguration :: Maybe AutoRollbackConfiguration

Configuration information for an automatic rollback that is added when a deployment group is created.

triggerConfigurations :: Maybe [TriggerConfig]

Information about triggers to create when the deployment group is created. For examples, see Create a Trigger for an AWS CodeDeploy Event in the AWS CodeDeploy User Guide.

ec2TagFilters :: Maybe [EC2TagFilter]

The Amazon EC2 tags on which to filter. The deployment group includes EC2 instances with any of the specified tags. Cannot be used in the same call as ec2TagSet.

onPremisesInstanceTagFilters :: Maybe [TagFilter]

The on-premises instance tags on which to filter. The deployment group includes on-premises instances with any of the specified tags. Cannot be used in the same call as OnPremisesTagSet.

blueGreenDeploymentConfiguration :: Maybe BlueGreenDeploymentConfiguration

Information about blue/green deployment options for a deployment group.

loadBalancerInfo :: Maybe LoadBalancerInfo

Information about the load balancer used in a deployment.

ec2TagSet :: Maybe EC2TagSet

Information about groups of tags applied to EC2 instances. The deployment group includes only EC2 instances identified by all the tag groups. Cannot be used in the same call as ec2TagFilters.

tags :: Maybe [Tag]

The metadata that you apply to CodeDeploy deployment groups to help you organize and categorize them. Each tag consists of a key and an optional value, both of which you define.

deploymentStyle :: Maybe DeploymentStyle

Information about the type of deployment, in-place or blue/green, that you want to run and whether to route deployment traffic behind a load balancer.

autoScalingGroups :: Maybe [Text]

A list of associated Amazon EC2 Auto Scaling groups.

ecsServices :: Maybe [ECSService]

The target Amazon ECS services in the deployment group. This applies only to deployment groups that use the Amazon ECS compute platform. A target Amazon ECS service is specified as an Amazon ECS cluster and service name pair using the format <clustername>:<servicename>.

alarmConfiguration :: Maybe AlarmConfiguration

Information to add about Amazon CloudWatch alarms when the deployment group is created.

applicationName :: Text

The name of an AWS CodeDeploy application associated with the IAM user or AWS account.

deploymentGroupName :: Text

The name of a new deployment group for the specified application.

serviceRoleArn :: Text

A service role Amazon Resource Name (ARN) that allows AWS CodeDeploy to act on the user's behalf when interacting with AWS services.

newCreateDeploymentGroup

Arguments

:: Text
$sel:applicationName:CreateDeploymentGroup'

-> Text
$sel:deploymentGroupName:CreateDeploymentGroup'

-> Text
$sel:serviceRoleArn:CreateDeploymentGroup'

-> CreateDeploymentGroup

Create a value of CreateDeploymentGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:onPremisesTagSet:CreateDeploymentGroup', createDeploymentGroup_onPremisesTagSet - Information about groups of tags applied to on-premises instances. The deployment group includes only on-premises instances identified by all of the tag groups. Cannot be used in the same call as onPremisesInstanceTagFilters.

$sel:outdatedInstancesStrategy:CreateDeploymentGroup', createDeploymentGroup_outdatedInstancesStrategy - Indicates what happens when new EC2 instances are launched mid-deployment and do not receive the deployed application revision.

If this option is set to UPDATE or is unspecified, CodeDeploy initiates one or more 'auto-update outdated instances' deployments to apply the deployed application revision to the new EC2 instances.

If this option is set to IGNORE, CodeDeploy does not initiate a deployment to update the new EC2 instances. This may result in instances having different revisions.

$sel:deploymentConfigName:CreateDeploymentGroup', createDeploymentGroup_deploymentConfigName - If specified, the deployment configuration name can be either one of the predefined configurations provided with AWS CodeDeploy or a custom deployment configuration that you create by calling the create deployment configuration operation.

CodeDeployDefault.OneAtATime is the default deployment configuration. It is used if a configuration isn't specified for the deployment or deployment group.

For more information about the predefined deployment configurations in AWS CodeDeploy, see Working with Deployment Configurations in CodeDeploy in the AWS CodeDeploy User Guide.

$sel:autoRollbackConfiguration:CreateDeploymentGroup', createDeploymentGroup_autoRollbackConfiguration - Configuration information for an automatic rollback that is added when a deployment group is created.

$sel:triggerConfigurations:CreateDeploymentGroup', createDeploymentGroup_triggerConfigurations - Information about triggers to create when the deployment group is created. For examples, see Create a Trigger for an AWS CodeDeploy Event in the AWS CodeDeploy User Guide.

$sel:ec2TagFilters:CreateDeploymentGroup', createDeploymentGroup_ec2TagFilters - The Amazon EC2 tags on which to filter. The deployment group includes EC2 instances with any of the specified tags. Cannot be used in the same call as ec2TagSet.

$sel:onPremisesInstanceTagFilters:CreateDeploymentGroup', createDeploymentGroup_onPremisesInstanceTagFilters - The on-premises instance tags on which to filter. The deployment group includes on-premises instances with any of the specified tags. Cannot be used in the same call as OnPremisesTagSet.

$sel:blueGreenDeploymentConfiguration:CreateDeploymentGroup', createDeploymentGroup_blueGreenDeploymentConfiguration - Information about blue/green deployment options for a deployment group.

$sel:loadBalancerInfo:CreateDeploymentGroup', createDeploymentGroup_loadBalancerInfo - Information about the load balancer used in a deployment.

$sel:ec2TagSet:CreateDeploymentGroup', createDeploymentGroup_ec2TagSet - Information about groups of tags applied to EC2 instances. The deployment group includes only EC2 instances identified by all the tag groups. Cannot be used in the same call as ec2TagFilters.

$sel:tags:CreateDeploymentGroup', createDeploymentGroup_tags - The metadata that you apply to CodeDeploy deployment groups to help you organize and categorize them. Each tag consists of a key and an optional value, both of which you define.

$sel:deploymentStyle:CreateDeploymentGroup', createDeploymentGroup_deploymentStyle - Information about the type of deployment, in-place or blue/green, that you want to run and whether to route deployment traffic behind a load balancer.

$sel:autoScalingGroups:CreateDeploymentGroup', createDeploymentGroup_autoScalingGroups - A list of associated Amazon EC2 Auto Scaling groups.

$sel:ecsServices:CreateDeploymentGroup', createDeploymentGroup_ecsServices - The target Amazon ECS services in the deployment group. This applies only to deployment groups that use the Amazon ECS compute platform. A target Amazon ECS service is specified as an Amazon ECS cluster and service name pair using the format <clustername>:<servicename>.

$sel:alarmConfiguration:CreateDeploymentGroup', createDeploymentGroup_alarmConfiguration - Information to add about Amazon CloudWatch alarms when the deployment group is created.

$sel:applicationName:CreateDeploymentGroup', createDeploymentGroup_applicationName - The name of an AWS CodeDeploy application associated with the IAM user or AWS account.

$sel:deploymentGroupName:CreateDeploymentGroup', createDeploymentGroup_deploymentGroupName - The name of a new deployment group for the specified application.

$sel:serviceRoleArn:CreateDeploymentGroup', createDeploymentGroup_serviceRoleArn - A service role Amazon Resource Name (ARN) that allows AWS CodeDeploy to act on the user's behalf when interacting with AWS services.

Request Lenses

createDeploymentGroup_onPremisesTagSet :: Lens' CreateDeploymentGroup (Maybe OnPremisesTagSet)

Information about groups of tags applied to on-premises instances. The deployment group includes only on-premises instances identified by all of the tag groups. Cannot be used in the same call as onPremisesInstanceTagFilters.

createDeploymentGroup_outdatedInstancesStrategy :: Lens' CreateDeploymentGroup (Maybe OutdatedInstancesStrategy)

Indicates what happens when new EC2 instances are launched mid-deployment and do not receive the deployed application revision.

If this option is set to UPDATE or is unspecified, CodeDeploy initiates one or more 'auto-update outdated instances' deployments to apply the deployed application revision to the new EC2 instances.

If this option is set to IGNORE, CodeDeploy does not initiate a deployment to update the new EC2 instances. This may result in instances having different revisions.

createDeploymentGroup_deploymentConfigName :: Lens' CreateDeploymentGroup (Maybe Text)

If specified, the deployment configuration name can be either one of the predefined configurations provided with AWS CodeDeploy or a custom deployment configuration that you create by calling the create deployment configuration operation.

CodeDeployDefault.OneAtATime is the default deployment configuration. It is used if a configuration isn't specified for the deployment or deployment group.

For more information about the predefined deployment configurations in AWS CodeDeploy, see Working with Deployment Configurations in CodeDeploy in the AWS CodeDeploy User Guide.

createDeploymentGroup_autoRollbackConfiguration :: Lens' CreateDeploymentGroup (Maybe AutoRollbackConfiguration)

Configuration information for an automatic rollback that is added when a deployment group is created.

createDeploymentGroup_triggerConfigurations :: Lens' CreateDeploymentGroup (Maybe [TriggerConfig])

Information about triggers to create when the deployment group is created. For examples, see Create a Trigger for an AWS CodeDeploy Event in the AWS CodeDeploy User Guide.

createDeploymentGroup_ec2TagFilters :: Lens' CreateDeploymentGroup (Maybe [EC2TagFilter])

The Amazon EC2 tags on which to filter. The deployment group includes EC2 instances with any of the specified tags. Cannot be used in the same call as ec2TagSet.

createDeploymentGroup_onPremisesInstanceTagFilters :: Lens' CreateDeploymentGroup (Maybe [TagFilter])

The on-premises instance tags on which to filter. The deployment group includes on-premises instances with any of the specified tags. Cannot be used in the same call as OnPremisesTagSet.

createDeploymentGroup_blueGreenDeploymentConfiguration :: Lens' CreateDeploymentGroup (Maybe BlueGreenDeploymentConfiguration)

Information about blue/green deployment options for a deployment group.

createDeploymentGroup_loadBalancerInfo :: Lens' CreateDeploymentGroup (Maybe LoadBalancerInfo)

Information about the load balancer used in a deployment.

createDeploymentGroup_ec2TagSet :: Lens' CreateDeploymentGroup (Maybe EC2TagSet)

Information about groups of tags applied to EC2 instances. The deployment group includes only EC2 instances identified by all the tag groups. Cannot be used in the same call as ec2TagFilters.

createDeploymentGroup_tags :: Lens' CreateDeploymentGroup (Maybe [Tag])

The metadata that you apply to CodeDeploy deployment groups to help you organize and categorize them. Each tag consists of a key and an optional value, both of which you define.

createDeploymentGroup_deploymentStyle :: Lens' CreateDeploymentGroup (Maybe DeploymentStyle)

Information about the type of deployment, in-place or blue/green, that you want to run and whether to route deployment traffic behind a load balancer.

createDeploymentGroup_autoScalingGroups :: Lens' CreateDeploymentGroup (Maybe [Text])

A list of associated Amazon EC2 Auto Scaling groups.

createDeploymentGroup_ecsServices :: Lens' CreateDeploymentGroup (Maybe [ECSService])

The target Amazon ECS services in the deployment group. This applies only to deployment groups that use the Amazon ECS compute platform. A target Amazon ECS service is specified as an Amazon ECS cluster and service name pair using the format <clustername>:<servicename>.

createDeploymentGroup_alarmConfiguration :: Lens' CreateDeploymentGroup (Maybe AlarmConfiguration)

Information to add about Amazon CloudWatch alarms when the deployment group is created.

createDeploymentGroup_applicationName :: Lens' CreateDeploymentGroup Text

The name of an AWS CodeDeploy application associated with the IAM user or AWS account.

createDeploymentGroup_deploymentGroupName :: Lens' CreateDeploymentGroup Text

The name of a new deployment group for the specified application.

createDeploymentGroup_serviceRoleArn :: Lens' CreateDeploymentGroup Text

A service role Amazon Resource Name (ARN) that allows AWS CodeDeploy to act on the user's behalf when interacting with AWS services.

Destructuring the Response

data CreateDeploymentGroupResponse

Represents the output of a CreateDeploymentGroup operation.

See: newCreateDeploymentGroupResponse smart constructor.

Constructors

CreateDeploymentGroupResponse'

Fields

deploymentGroupId :: Maybe Text

A unique deployment group ID.

httpStatus :: Int

The response's http status code.

newCreateDeploymentGroupResponse

Arguments

:: Int
$sel:httpStatus:CreateDeploymentGroupResponse'

-> CreateDeploymentGroupResponse

Create a value of CreateDeploymentGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deploymentGroupId:CreateDeploymentGroupResponse', createDeploymentGroupResponse_deploymentGroupId - A unique deployment group ID.

$sel:httpStatus:CreateDeploymentGroupResponse', createDeploymentGroupResponse_httpStatus - The response's http status code.

Response Lenses

createDeploymentGroupResponse_deploymentGroupId :: Lens' CreateDeploymentGroupResponse (Maybe Text)

A unique deployment group ID.

createDeploymentGroupResponse_httpStatus :: Lens' CreateDeploymentGroupResponse Int

The response's http status code.