Network.AWS.CodeDeploy.GetOnPremisesInstance

Description

Gets information about an on-premises instance.

Creating a Request

data GetOnPremisesInstance

Represents the input of a GetOnPremisesInstance operation.

See: newGetOnPremisesInstance smart constructor.

Constructors

GetOnPremisesInstance'

Fields

instanceName :: Text

The name of the on-premises instance about which to get information.

newGetOnPremisesInstance

Arguments

:: Text
$sel:instanceName:GetOnPremisesInstance'

-> GetOnPremisesInstance

Create a value of GetOnPremisesInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceName:GetOnPremisesInstance', getOnPremisesInstance_instanceName - The name of the on-premises instance about which to get information.

Request Lenses

getOnPremisesInstance_instanceName :: Lens' GetOnPremisesInstance Text

The name of the on-premises instance about which to get information.

Destructuring the Response

data GetOnPremisesInstanceResponse

Represents the output of a GetOnPremisesInstance operation.

See: newGetOnPremisesInstanceResponse smart constructor.

Constructors

GetOnPremisesInstanceResponse'

Fields

instanceInfo :: Maybe InstanceInfo

Information about the on-premises instance.

httpStatus :: Int

The response's http status code.

newGetOnPremisesInstanceResponse

Arguments

:: Int
$sel:httpStatus:GetOnPremisesInstanceResponse'

-> GetOnPremisesInstanceResponse

Create a value of GetOnPremisesInstanceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceInfo:GetOnPremisesInstanceResponse', getOnPremisesInstanceResponse_instanceInfo - Information about the on-premises instance.

$sel:httpStatus:GetOnPremisesInstanceResponse', getOnPremisesInstanceResponse_httpStatus - The response's http status code.

Response Lenses

getOnPremisesInstanceResponse_instanceInfo :: Lens' GetOnPremisesInstanceResponse (Maybe InstanceInfo)

Information about the on-premises instance.

getOnPremisesInstanceResponse_httpStatus :: Lens' GetOnPremisesInstanceResponse Int

The response's http status code.