Network.AWS.CodeDeploy.ListDeploymentTargets

Description

Returns an array of target IDs that are associated a deployment.

This operation returns paginated results.

Creating a Request

data ListDeploymentTargets

See: newListDeploymentTargets smart constructor.

Constructors

ListDeploymentTargets'

Fields

deploymentId :: Maybe Text

The unique ID of a deployment.

nextToken :: Maybe Text

A token identifier returned from the previous ListDeploymentTargets call. It can be used to return the next set of deployment targets in the list.

targetFilters :: Maybe (HashMap TargetFilterName [Text])

A key used to filter the returned targets. The two valid values are:

TargetStatus - A TargetStatus filter string can be Failed, InProgress, Pending, Ready, Skipped, Succeeded, or Unknown.

ServerInstanceLabel - A ServerInstanceLabel filter string can be Blue or Green.

newListDeploymentTargets :: ListDeploymentTargets

Create a value of ListDeploymentTargets with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deploymentId:ListDeploymentTargets', listDeploymentTargets_deploymentId - The unique ID of a deployment.

$sel:nextToken:ListDeploymentTargets', listDeploymentTargets_nextToken - A token identifier returned from the previous ListDeploymentTargets call. It can be used to return the next set of deployment targets in the list.

$sel:targetFilters:ListDeploymentTargets', listDeploymentTargets_targetFilters - A key used to filter the returned targets. The two valid values are:

TargetStatus - A TargetStatus filter string can be Failed, InProgress, Pending, Ready, Skipped, Succeeded, or Unknown.

ServerInstanceLabel - A ServerInstanceLabel filter string can be Blue or Green.

Request Lenses

listDeploymentTargets_deploymentId :: Lens' ListDeploymentTargets (Maybe Text)

The unique ID of a deployment.

listDeploymentTargets_nextToken :: Lens' ListDeploymentTargets (Maybe Text)

A token identifier returned from the previous ListDeploymentTargets call. It can be used to return the next set of deployment targets in the list.

listDeploymentTargets_targetFilters :: Lens' ListDeploymentTargets (Maybe (HashMap TargetFilterName [Text]))

A key used to filter the returned targets. The two valid values are:

TargetStatus - A TargetStatus filter string can be Failed, InProgress, Pending, Ready, Skipped, Succeeded, or Unknown.

ServerInstanceLabel - A ServerInstanceLabel filter string can be Blue or Green.

Destructuring the Response

data ListDeploymentTargetsResponse

See: newListDeploymentTargetsResponse smart constructor.

Constructors

ListDeploymentTargetsResponse'

Fields

targetIds :: Maybe [Text]

The unique IDs of deployment targets.

nextToken :: Maybe Text

If a large amount of information is returned, a token identifier is also returned. It can be used in a subsequent ListDeploymentTargets call to return the next set of deployment targets in the list.

httpStatus :: Int

The response's http status code.

newListDeploymentTargetsResponse

Arguments

:: Int
$sel:httpStatus:ListDeploymentTargetsResponse'

-> ListDeploymentTargetsResponse

Create a value of ListDeploymentTargetsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetIds:ListDeploymentTargetsResponse', listDeploymentTargetsResponse_targetIds - The unique IDs of deployment targets.

$sel:nextToken:ListDeploymentTargets', listDeploymentTargetsResponse_nextToken - If a large amount of information is returned, a token identifier is also returned. It can be used in a subsequent ListDeploymentTargets call to return the next set of deployment targets in the list.

$sel:httpStatus:ListDeploymentTargetsResponse', listDeploymentTargetsResponse_httpStatus - The response's http status code.

Response Lenses

listDeploymentTargetsResponse_targetIds :: Lens' ListDeploymentTargetsResponse (Maybe [Text])

The unique IDs of deployment targets.

listDeploymentTargetsResponse_nextToken :: Lens' ListDeploymentTargetsResponse (Maybe Text)

If a large amount of information is returned, a token identifier is also returned. It can be used in a subsequent ListDeploymentTargets call to return the next set of deployment targets in the list.

listDeploymentTargetsResponse_httpStatus :: Lens' ListDeploymentTargetsResponse Int

The response's http status code.