Network.AWS.CodeDeploy.ListDeployments

Description

Lists the deployments in a deployment group for an application registered with the IAM user or AWS account.

This operation returns paginated results.

Creating a Request

data ListDeployments

Represents the input of a ListDeployments operation.

See: newListDeployments smart constructor.

Constructors

ListDeployments'

Fields

nextToken :: Maybe Text

An identifier returned from the previous list deployments call. It can be used to return the next set of deployments in the list.

deploymentGroupName :: Maybe Text

The name of a deployment group for the specified application.

If deploymentGroupName is specified, then applicationName must be specified. If it is not specified, then applicationName must not be specified.

createTimeRange :: Maybe TimeRange

A time range (start and end) for returning a subset of the list of deployments.

includeOnlyStatuses :: Maybe [DeploymentStatus]

A subset of deployments to list by status:

Created: Include created deployments in the resulting list.

Queued: Include queued deployments in the resulting list.

In Progress: Include in-progress deployments in the resulting list.

Succeeded: Include successful deployments in the resulting list.

Failed: Include failed deployments in the resulting list.

Stopped: Include stopped deployments in the resulting list.

externalId :: Maybe Text

The unique ID of an external resource for returning deployments linked to the external resource.

applicationName :: Maybe Text

The name of an AWS CodeDeploy application associated with the IAM user or AWS account.

If applicationName is specified, then deploymentGroupName must be specified. If it is not specified, then deploymentGroupName must not be specified.

newListDeployments :: ListDeployments

Create a value of ListDeployments with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDeployments', listDeployments_nextToken - An identifier returned from the previous list deployments call. It can be used to return the next set of deployments in the list.

$sel:deploymentGroupName:ListDeployments', listDeployments_deploymentGroupName - The name of a deployment group for the specified application.

If deploymentGroupName is specified, then applicationName must be specified. If it is not specified, then applicationName must not be specified.

$sel:createTimeRange:ListDeployments', listDeployments_createTimeRange - A time range (start and end) for returning a subset of the list of deployments.

$sel:includeOnlyStatuses:ListDeployments', listDeployments_includeOnlyStatuses - A subset of deployments to list by status:

Created: Include created deployments in the resulting list.

Queued: Include queued deployments in the resulting list.

In Progress: Include in-progress deployments in the resulting list.

Succeeded: Include successful deployments in the resulting list.

Failed: Include failed deployments in the resulting list.

Stopped: Include stopped deployments in the resulting list.

$sel:externalId:ListDeployments', listDeployments_externalId - The unique ID of an external resource for returning deployments linked to the external resource.

$sel:applicationName:ListDeployments', listDeployments_applicationName - The name of an AWS CodeDeploy application associated with the IAM user or AWS account.

If applicationName is specified, then deploymentGroupName must be specified. If it is not specified, then deploymentGroupName must not be specified.

Request Lenses

listDeployments_nextToken :: Lens' ListDeployments (Maybe Text)

An identifier returned from the previous list deployments call. It can be used to return the next set of deployments in the list.

listDeployments_deploymentGroupName :: Lens' ListDeployments (Maybe Text)

The name of a deployment group for the specified application.

If deploymentGroupName is specified, then applicationName must be specified. If it is not specified, then applicationName must not be specified.

listDeployments_createTimeRange :: Lens' ListDeployments (Maybe TimeRange)

A time range (start and end) for returning a subset of the list of deployments.

listDeployments_includeOnlyStatuses :: Lens' ListDeployments (Maybe [DeploymentStatus])

A subset of deployments to list by status:

Created: Include created deployments in the resulting list.

Queued: Include queued deployments in the resulting list.

In Progress: Include in-progress deployments in the resulting list.

Succeeded: Include successful deployments in the resulting list.

Failed: Include failed deployments in the resulting list.

Stopped: Include stopped deployments in the resulting list.

listDeployments_externalId :: Lens' ListDeployments (Maybe Text)

The unique ID of an external resource for returning deployments linked to the external resource.

listDeployments_applicationName :: Lens' ListDeployments (Maybe Text)

The name of an AWS CodeDeploy application associated with the IAM user or AWS account.

If applicationName is specified, then deploymentGroupName must be specified. If it is not specified, then deploymentGroupName must not be specified.

Destructuring the Response

data ListDeploymentsResponse

Represents the output of a ListDeployments operation.

See: newListDeploymentsResponse smart constructor.

Constructors

ListDeploymentsResponse'

Fields

nextToken :: Maybe Text

If a large amount of information is returned, an identifier is also returned. It can be used in a subsequent list deployments call to return the next set of deployments in the list.

deployments :: Maybe [Text]

A list of deployment IDs.

httpStatus :: Int

The response's http status code.

newListDeploymentsResponse

Arguments

:: Int
$sel:httpStatus:ListDeploymentsResponse'

-> ListDeploymentsResponse

Create a value of ListDeploymentsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDeployments', listDeploymentsResponse_nextToken - If a large amount of information is returned, an identifier is also returned. It can be used in a subsequent list deployments call to return the next set of deployments in the list.

$sel:deployments:ListDeploymentsResponse', listDeploymentsResponse_deployments - A list of deployment IDs.

$sel:httpStatus:ListDeploymentsResponse', listDeploymentsResponse_httpStatus - The response's http status code.

Response Lenses

listDeploymentsResponse_nextToken :: Lens' ListDeploymentsResponse (Maybe Text)

If a large amount of information is returned, an identifier is also returned. It can be used in a subsequent list deployments call to return the next set of deployments in the list.

listDeploymentsResponse_deployments :: Lens' ListDeploymentsResponse (Maybe [Text])

A list of deployment IDs.

listDeploymentsResponse_httpStatus :: Lens' ListDeploymentsResponse Int

The response's http status code.