Network.AWS.CodeDeploy.ListOnPremisesInstances

Description

Gets a list of names for one or more on-premises instances.

Unless otherwise specified, both registered and deregistered on-premises instance names are listed. To list only registered or deregistered on-premises instance names, use the registration status parameter.

This operation returns paginated results.

Creating a Request

data ListOnPremisesInstances

Represents the input of a ListOnPremisesInstances operation.

See: newListOnPremisesInstances smart constructor.

Constructors

ListOnPremisesInstances'

Fields

nextToken :: Maybe Text

An identifier returned from the previous list on-premises instances call. It can be used to return the next set of on-premises instances in the list.

tagFilters :: Maybe [TagFilter]

The on-premises instance tags that are used to restrict the on-premises instance names returned.

registrationStatus :: Maybe RegistrationStatus

The registration status of the on-premises instances:

Deregistered: Include deregistered on-premises instances in the resulting list.

Registered: Include registered on-premises instances in the resulting list.

newListOnPremisesInstances :: ListOnPremisesInstances

Create a value of ListOnPremisesInstances with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListOnPremisesInstances', listOnPremisesInstances_nextToken - An identifier returned from the previous list on-premises instances call. It can be used to return the next set of on-premises instances in the list.

$sel:tagFilters:ListOnPremisesInstances', listOnPremisesInstances_tagFilters - The on-premises instance tags that are used to restrict the on-premises instance names returned.

$sel:registrationStatus:ListOnPremisesInstances', listOnPremisesInstances_registrationStatus - The registration status of the on-premises instances:

Deregistered: Include deregistered on-premises instances in the resulting list.

Registered: Include registered on-premises instances in the resulting list.

Request Lenses

listOnPremisesInstances_nextToken :: Lens' ListOnPremisesInstances (Maybe Text)

An identifier returned from the previous list on-premises instances call. It can be used to return the next set of on-premises instances in the list.

listOnPremisesInstances_tagFilters :: Lens' ListOnPremisesInstances (Maybe [TagFilter])

The on-premises instance tags that are used to restrict the on-premises instance names returned.

listOnPremisesInstances_registrationStatus :: Lens' ListOnPremisesInstances (Maybe RegistrationStatus)

The registration status of the on-premises instances:

Deregistered: Include deregistered on-premises instances in the resulting list.

Registered: Include registered on-premises instances in the resulting list.

Destructuring the Response

data ListOnPremisesInstancesResponse

Represents the output of the list on-premises instances operation.

See: newListOnPremisesInstancesResponse smart constructor.

Constructors

ListOnPremisesInstancesResponse'

Fields

nextToken :: Maybe Text

If a large amount of information is returned, an identifier is also returned. It can be used in a subsequent list on-premises instances call to return the next set of on-premises instances in the list.

instanceNames :: Maybe [Text]

The list of matching on-premises instance names.

httpStatus :: Int

The response's http status code.

newListOnPremisesInstancesResponse

Arguments

:: Int
$sel:httpStatus:ListOnPremisesInstancesResponse'

-> ListOnPremisesInstancesResponse

Create a value of ListOnPremisesInstancesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListOnPremisesInstances', listOnPremisesInstancesResponse_nextToken - If a large amount of information is returned, an identifier is also returned. It can be used in a subsequent list on-premises instances call to return the next set of on-premises instances in the list.

$sel:instanceNames:ListOnPremisesInstancesResponse', listOnPremisesInstancesResponse_instanceNames - The list of matching on-premises instance names.

$sel:httpStatus:ListOnPremisesInstancesResponse', listOnPremisesInstancesResponse_httpStatus - The response's http status code.

Response Lenses

listOnPremisesInstancesResponse_nextToken :: Lens' ListOnPremisesInstancesResponse (Maybe Text)

If a large amount of information is returned, an identifier is also returned. It can be used in a subsequent list on-premises instances call to return the next set of on-premises instances in the list.

listOnPremisesInstancesResponse_instanceNames :: Lens' ListOnPremisesInstancesResponse (Maybe [Text])

The list of matching on-premises instance names.

listOnPremisesInstancesResponse_httpStatus :: Lens' ListOnPremisesInstancesResponse Int

The response's http status code.