Network.AWS.CodeDeploy.PutLifecycleEventHookExecutionStatus

Description

Sets the result of a Lambda validation function. The function validates lifecycle hooks during a deployment that uses the AWS Lambda or Amazon ECS compute platform. For AWS Lambda deployments, the available lifecycle hooks are BeforeAllowTraffic and AfterAllowTraffic. For Amazon ECS deployments, the available lifecycle hooks are BeforeInstall, AfterInstall, AfterAllowTestTraffic, BeforeAllowTraffic, and AfterAllowTraffic. Lambda validation functions return Succeeded or Failed. For more information, see AppSpec 'hooks' Section for an AWS Lambda Deployment and AppSpec 'hooks' Section for an Amazon ECS Deployment.

Creating a Request

data PutLifecycleEventHookExecutionStatus

See: newPutLifecycleEventHookExecutionStatus smart constructor.

Constructors

PutLifecycleEventHookExecutionStatus'

Fields

status :: Maybe LifecycleEventStatus

The result of a Lambda function that validates a deployment lifecycle event. Succeeded and Failed are the only valid values for status.

deploymentId :: Maybe Text

The unique ID of a deployment. Pass this ID to a Lambda function that validates a deployment lifecycle event.

lifecycleEventHookExecutionId :: Maybe Text

The execution ID of a deployment's lifecycle hook. A deployment lifecycle hook is specified in the hooks section of the AppSpec file.

newPutLifecycleEventHookExecutionStatus :: PutLifecycleEventHookExecutionStatus

Create a value of PutLifecycleEventHookExecutionStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:PutLifecycleEventHookExecutionStatus', putLifecycleEventHookExecutionStatus_status - The result of a Lambda function that validates a deployment lifecycle event. Succeeded and Failed are the only valid values for status.

$sel:deploymentId:PutLifecycleEventHookExecutionStatus', putLifecycleEventHookExecutionStatus_deploymentId - The unique ID of a deployment. Pass this ID to a Lambda function that validates a deployment lifecycle event.

$sel:lifecycleEventHookExecutionId:PutLifecycleEventHookExecutionStatus', putLifecycleEventHookExecutionStatus_lifecycleEventHookExecutionId - The execution ID of a deployment's lifecycle hook. A deployment lifecycle hook is specified in the hooks section of the AppSpec file.

Request Lenses

putLifecycleEventHookExecutionStatus_status :: Lens' PutLifecycleEventHookExecutionStatus (Maybe LifecycleEventStatus)

The result of a Lambda function that validates a deployment lifecycle event. Succeeded and Failed are the only valid values for status.

putLifecycleEventHookExecutionStatus_deploymentId :: Lens' PutLifecycleEventHookExecutionStatus (Maybe Text)

The unique ID of a deployment. Pass this ID to a Lambda function that validates a deployment lifecycle event.

putLifecycleEventHookExecutionStatus_lifecycleEventHookExecutionId :: Lens' PutLifecycleEventHookExecutionStatus (Maybe Text)

The execution ID of a deployment's lifecycle hook. A deployment lifecycle hook is specified in the hooks section of the AppSpec file.

Destructuring the Response

data PutLifecycleEventHookExecutionStatusResponse

See: newPutLifecycleEventHookExecutionStatusResponse smart constructor.

Constructors

PutLifecycleEventHookExecutionStatusResponse'

Fields

lifecycleEventHookExecutionId :: Maybe Text

The execution ID of the lifecycle event hook. A hook is specified in the hooks section of the deployment's AppSpec file.

httpStatus :: Int

The response's http status code.

newPutLifecycleEventHookExecutionStatusResponse

Arguments

:: Int
$sel:httpStatus:PutLifecycleEventHookExecutionStatusResponse'

-> PutLifecycleEventHookExecutionStatusResponse

Create a value of PutLifecycleEventHookExecutionStatusResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lifecycleEventHookExecutionId:PutLifecycleEventHookExecutionStatus', putLifecycleEventHookExecutionStatusResponse_lifecycleEventHookExecutionId - The execution ID of the lifecycle event hook. A hook is specified in the hooks section of the deployment's AppSpec file.

$sel:httpStatus:PutLifecycleEventHookExecutionStatusResponse', putLifecycleEventHookExecutionStatusResponse_httpStatus - The response's http status code.

Response Lenses

putLifecycleEventHookExecutionStatusResponse_lifecycleEventHookExecutionId :: Lens' PutLifecycleEventHookExecutionStatusResponse (Maybe Text)

The execution ID of the lifecycle event hook. A hook is specified in the hooks section of the deployment's AppSpec file.

putLifecycleEventHookExecutionStatusResponse_httpStatus :: Lens' PutLifecycleEventHookExecutionStatusResponse Int

The response's http status code.