Network.AWS.CodeDeploy.StopDeployment

Description

Attempts to stop an ongoing deployment.

Creating a Request

data StopDeployment

Represents the input of a StopDeployment operation.

See: newStopDeployment smart constructor.

Constructors

StopDeployment'

Fields

autoRollbackEnabled :: Maybe Bool

Indicates, when a deployment is stopped, whether instances that have been updated should be rolled back to the previous version of the application revision.

deploymentId :: Text

The unique ID of a deployment.

newStopDeployment

Arguments

:: Text
$sel:deploymentId:StopDeployment'

-> StopDeployment

Create a value of StopDeployment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:autoRollbackEnabled:StopDeployment', stopDeployment_autoRollbackEnabled - Indicates, when a deployment is stopped, whether instances that have been updated should be rolled back to the previous version of the application revision.

$sel:deploymentId:StopDeployment', stopDeployment_deploymentId - The unique ID of a deployment.

Request Lenses

stopDeployment_autoRollbackEnabled :: Lens' StopDeployment (Maybe Bool)

Indicates, when a deployment is stopped, whether instances that have been updated should be rolled back to the previous version of the application revision.

stopDeployment_deploymentId :: Lens' StopDeployment Text

The unique ID of a deployment.

Destructuring the Response

data StopDeploymentResponse

Represents the output of a StopDeployment operation.

See: newStopDeploymentResponse smart constructor.

Constructors

StopDeploymentResponse'

Fields

statusMessage :: Maybe Text

An accompanying status message.

status :: Maybe StopStatus

The status of the stop deployment operation:

Pending: The stop operation is pending.

Succeeded: The stop operation was successful.

httpStatus :: Int

The response's http status code.

newStopDeploymentResponse

Arguments

:: Int
$sel:httpStatus:StopDeploymentResponse'

-> StopDeploymentResponse

Create a value of StopDeploymentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:statusMessage:StopDeploymentResponse', stopDeploymentResponse_statusMessage - An accompanying status message.

$sel:status:StopDeploymentResponse', stopDeploymentResponse_status - The status of the stop deployment operation:

Pending: The stop operation is pending.

Succeeded: The stop operation was successful.

$sel:httpStatus:StopDeploymentResponse', stopDeploymentResponse_httpStatus - The response's http status code.

Response Lenses

stopDeploymentResponse_statusMessage :: Lens' StopDeploymentResponse (Maybe Text)

An accompanying status message.

stopDeploymentResponse_status :: Lens' StopDeploymentResponse (Maybe StopStatus)

The status of the stop deployment operation:

Pending: The stop operation is pending.

Succeeded: The stop operation was successful.

stopDeploymentResponse_httpStatus :: Lens' StopDeploymentResponse Int

The response's http status code.