Network.AWS.CodeDeploy.Types.AlarmConfiguration

Description

Documentation

data AlarmConfiguration

Information about alarms associated with the deployment group.

See: newAlarmConfiguration smart constructor.

Constructors

AlarmConfiguration'

Fields

ignorePollAlarmFailure :: Maybe Bool

Indicates whether a deployment should continue if information about the current state of alarms cannot be retrieved from Amazon CloudWatch. The default value is false.

true: The deployment proceeds even if alarm status information can't be retrieved from Amazon CloudWatch.

false: The deployment stops if alarm status information can't be retrieved from Amazon CloudWatch.

enabled :: Maybe Bool

Indicates whether the alarm configuration is enabled.

alarms :: Maybe [Alarm]

A list of alarms configured for the deployment group. A maximum of 10 alarms can be added to a deployment group.

newAlarmConfiguration :: AlarmConfiguration

Create a value of AlarmConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ignorePollAlarmFailure:AlarmConfiguration', alarmConfiguration_ignorePollAlarmFailure - Indicates whether a deployment should continue if information about the current state of alarms cannot be retrieved from Amazon CloudWatch. The default value is false.

true: The deployment proceeds even if alarm status information can't be retrieved from Amazon CloudWatch.

false: The deployment stops if alarm status information can't be retrieved from Amazon CloudWatch.

$sel:enabled:AlarmConfiguration', alarmConfiguration_enabled - Indicates whether the alarm configuration is enabled.

$sel:alarms:AlarmConfiguration', alarmConfiguration_alarms - A list of alarms configured for the deployment group. A maximum of 10 alarms can be added to a deployment group.

alarmConfiguration_ignorePollAlarmFailure :: Lens' AlarmConfiguration (Maybe Bool)

Indicates whether a deployment should continue if information about the current state of alarms cannot be retrieved from Amazon CloudWatch. The default value is false.

true: The deployment proceeds even if alarm status information can't be retrieved from Amazon CloudWatch.

false: The deployment stops if alarm status information can't be retrieved from Amazon CloudWatch.

alarmConfiguration_enabled :: Lens' AlarmConfiguration (Maybe Bool)

Indicates whether the alarm configuration is enabled.

alarmConfiguration_alarms :: Lens' AlarmConfiguration (Maybe [Alarm])

A list of alarms configured for the deployment group. A maximum of 10 alarms can be added to a deployment group.