Network.AWS.CodeDeploy.Types.BlueGreenDeploymentConfiguration

Description

Documentation

data BlueGreenDeploymentConfiguration

Information about blue/green deployment options for a deployment group.

See: newBlueGreenDeploymentConfiguration smart constructor.

Constructors

BlueGreenDeploymentConfiguration'

Fields

greenFleetProvisioningOption :: Maybe GreenFleetProvisioningOption

Information about how instances are provisioned for a replacement environment in a blue/green deployment.

deploymentReadyOption :: Maybe DeploymentReadyOption

Information about the action to take when newly provisioned instances are ready to receive traffic in a blue/green deployment.

terminateBlueInstancesOnDeploymentSuccess :: Maybe BlueInstanceTerminationOption

Information about whether to terminate instances in the original fleet during a blue/green deployment.

newBlueGreenDeploymentConfiguration :: BlueGreenDeploymentConfiguration

Create a value of BlueGreenDeploymentConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:greenFleetProvisioningOption:BlueGreenDeploymentConfiguration', blueGreenDeploymentConfiguration_greenFleetProvisioningOption - Information about how instances are provisioned for a replacement environment in a blue/green deployment.

$sel:deploymentReadyOption:BlueGreenDeploymentConfiguration', blueGreenDeploymentConfiguration_deploymentReadyOption - Information about the action to take when newly provisioned instances are ready to receive traffic in a blue/green deployment.

$sel:terminateBlueInstancesOnDeploymentSuccess:BlueGreenDeploymentConfiguration', blueGreenDeploymentConfiguration_terminateBlueInstancesOnDeploymentSuccess - Information about whether to terminate instances in the original fleet during a blue/green deployment.

blueGreenDeploymentConfiguration_greenFleetProvisioningOption :: Lens' BlueGreenDeploymentConfiguration (Maybe GreenFleetProvisioningOption)

Information about how instances are provisioned for a replacement environment in a blue/green deployment.

blueGreenDeploymentConfiguration_deploymentReadyOption :: Lens' BlueGreenDeploymentConfiguration (Maybe DeploymentReadyOption)

Information about the action to take when newly provisioned instances are ready to receive traffic in a blue/green deployment.

blueGreenDeploymentConfiguration_terminateBlueInstancesOnDeploymentSuccess :: Lens' BlueGreenDeploymentConfiguration (Maybe BlueInstanceTerminationOption)

Information about whether to terminate instances in the original fleet during a blue/green deployment.