Network.AWS.CodeDeploy.Types.DeploymentGroupInfo

Description

Documentation

data DeploymentGroupInfo

Information about a deployment group.

See: newDeploymentGroupInfo smart constructor.

Constructors

DeploymentGroupInfo'

Fields

onPremisesTagSet :: Maybe OnPremisesTagSet

Information about groups of tags applied to an on-premises instance. The deployment group includes only on-premises instances identified by all the tag groups. Cannot be used in the same call as onPremisesInstanceTagFilters.

outdatedInstancesStrategy :: Maybe OutdatedInstancesStrategy

Indicates what happens when new EC2 instances are launched mid-deployment and do not receive the deployed application revision.

If this option is set to UPDATE or is unspecified, CodeDeploy initiates one or more 'auto-update outdated instances' deployments to apply the deployed application revision to the new EC2 instances.

If this option is set to IGNORE, CodeDeploy does not initiate a deployment to update the new EC2 instances. This may result in instances having different revisions.

serviceRoleArn :: Maybe Text

A service role Amazon Resource Name (ARN) that grants CodeDeploy permission to make calls to AWS services on your behalf. For more information, see Create a Service Role for AWS CodeDeploy in the AWS CodeDeploy User Guide.

deploymentConfigName :: Maybe Text

The deployment configuration name.

deploymentGroupName :: Maybe Text

The deployment group name.

autoRollbackConfiguration :: Maybe AutoRollbackConfiguration

Information about the automatic rollback configuration associated with the deployment group.

deploymentGroupId :: Maybe Text

The deployment group ID.

triggerConfigurations :: Maybe [TriggerConfig]

Information about triggers associated with the deployment group.

ec2TagFilters :: Maybe [EC2TagFilter]

The Amazon EC2 tags on which to filter. The deployment group includes EC2 instances with any of the specified tags.

targetRevision :: Maybe RevisionLocation

Information about the deployment group's target revision, including type and location.

lastAttemptedDeployment :: Maybe LastDeploymentInfo

Information about the most recent attempted deployment to the deployment group.

lastSuccessfulDeployment :: Maybe LastDeploymentInfo

Information about the most recent successful deployment to the deployment group.

onPremisesInstanceTagFilters :: Maybe [TagFilter]

The on-premises instance tags on which to filter. The deployment group includes on-premises instances with any of the specified tags.

blueGreenDeploymentConfiguration :: Maybe BlueGreenDeploymentConfiguration

Information about blue/green deployment options for a deployment group.

loadBalancerInfo :: Maybe LoadBalancerInfo

Information about the load balancer to use in a deployment.

ec2TagSet :: Maybe EC2TagSet

Information about groups of tags applied to an EC2 instance. The deployment group includes only EC2 instances identified by all of the tag groups. Cannot be used in the same call as ec2TagFilters.

deploymentStyle :: Maybe DeploymentStyle

Information about the type of deployment, either in-place or blue/green, you want to run and whether to route deployment traffic behind a load balancer.

autoScalingGroups :: Maybe [AutoScalingGroup]

A list of associated Auto Scaling groups.

ecsServices :: Maybe [ECSService]

The target Amazon ECS services in the deployment group. This applies only to deployment groups that use the Amazon ECS compute platform. A target Amazon ECS service is specified as an Amazon ECS cluster and service name pair using the format <clustername>:<servicename>.

alarmConfiguration :: Maybe AlarmConfiguration

A list of alarms associated with the deployment group.

applicationName :: Maybe Text

The application name.

computePlatform :: Maybe ComputePlatform

The destination platform type for the deployment (Lambda, Server, or ECS).

newDeploymentGroupInfo :: DeploymentGroupInfo

Create a value of DeploymentGroupInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:onPremisesTagSet:DeploymentGroupInfo', deploymentGroupInfo_onPremisesTagSet - Information about groups of tags applied to an on-premises instance. The deployment group includes only on-premises instances identified by all the tag groups. Cannot be used in the same call as onPremisesInstanceTagFilters.

$sel:outdatedInstancesStrategy:DeploymentGroupInfo', deploymentGroupInfo_outdatedInstancesStrategy - Indicates what happens when new EC2 instances are launched mid-deployment and do not receive the deployed application revision.

If this option is set to UPDATE or is unspecified, CodeDeploy initiates one or more 'auto-update outdated instances' deployments to apply the deployed application revision to the new EC2 instances.

If this option is set to IGNORE, CodeDeploy does not initiate a deployment to update the new EC2 instances. This may result in instances having different revisions.

$sel:serviceRoleArn:DeploymentGroupInfo', deploymentGroupInfo_serviceRoleArn - A service role Amazon Resource Name (ARN) that grants CodeDeploy permission to make calls to AWS services on your behalf. For more information, see Create a Service Role for AWS CodeDeploy in the AWS CodeDeploy User Guide.

$sel:deploymentConfigName:DeploymentGroupInfo', deploymentGroupInfo_deploymentConfigName - The deployment configuration name.

$sel:deploymentGroupName:DeploymentGroupInfo', deploymentGroupInfo_deploymentGroupName - The deployment group name.

$sel:autoRollbackConfiguration:DeploymentGroupInfo', deploymentGroupInfo_autoRollbackConfiguration - Information about the automatic rollback configuration associated with the deployment group.

$sel:deploymentGroupId:DeploymentGroupInfo', deploymentGroupInfo_deploymentGroupId - The deployment group ID.

$sel:triggerConfigurations:DeploymentGroupInfo', deploymentGroupInfo_triggerConfigurations - Information about triggers associated with the deployment group.

$sel:ec2TagFilters:DeploymentGroupInfo', deploymentGroupInfo_ec2TagFilters - The Amazon EC2 tags on which to filter. The deployment group includes EC2 instances with any of the specified tags.

$sel:targetRevision:DeploymentGroupInfo', deploymentGroupInfo_targetRevision - Information about the deployment group's target revision, including type and location.

$sel:lastAttemptedDeployment:DeploymentGroupInfo', deploymentGroupInfo_lastAttemptedDeployment - Information about the most recent attempted deployment to the deployment group.

$sel:lastSuccessfulDeployment:DeploymentGroupInfo', deploymentGroupInfo_lastSuccessfulDeployment - Information about the most recent successful deployment to the deployment group.

$sel:onPremisesInstanceTagFilters:DeploymentGroupInfo', deploymentGroupInfo_onPremisesInstanceTagFilters - The on-premises instance tags on which to filter. The deployment group includes on-premises instances with any of the specified tags.

$sel:blueGreenDeploymentConfiguration:DeploymentGroupInfo', deploymentGroupInfo_blueGreenDeploymentConfiguration - Information about blue/green deployment options for a deployment group.

$sel:loadBalancerInfo:DeploymentGroupInfo', deploymentGroupInfo_loadBalancerInfo - Information about the load balancer to use in a deployment.

$sel:ec2TagSet:DeploymentGroupInfo', deploymentGroupInfo_ec2TagSet - Information about groups of tags applied to an EC2 instance. The deployment group includes only EC2 instances identified by all of the tag groups. Cannot be used in the same call as ec2TagFilters.

$sel:deploymentStyle:DeploymentGroupInfo', deploymentGroupInfo_deploymentStyle - Information about the type of deployment, either in-place or blue/green, you want to run and whether to route deployment traffic behind a load balancer.

$sel:autoScalingGroups:DeploymentGroupInfo', deploymentGroupInfo_autoScalingGroups - A list of associated Auto Scaling groups.

$sel:ecsServices:DeploymentGroupInfo', deploymentGroupInfo_ecsServices - The target Amazon ECS services in the deployment group. This applies only to deployment groups that use the Amazon ECS compute platform. A target Amazon ECS service is specified as an Amazon ECS cluster and service name pair using the format <clustername>:<servicename>.

$sel:alarmConfiguration:DeploymentGroupInfo', deploymentGroupInfo_alarmConfiguration - A list of alarms associated with the deployment group.

$sel:applicationName:DeploymentGroupInfo', deploymentGroupInfo_applicationName - The application name.

$sel:computePlatform:DeploymentGroupInfo', deploymentGroupInfo_computePlatform - The destination platform type for the deployment (Lambda, Server, or ECS).

deploymentGroupInfo_onPremisesTagSet :: Lens' DeploymentGroupInfo (Maybe OnPremisesTagSet)

Information about groups of tags applied to an on-premises instance. The deployment group includes only on-premises instances identified by all the tag groups. Cannot be used in the same call as onPremisesInstanceTagFilters.

deploymentGroupInfo_outdatedInstancesStrategy :: Lens' DeploymentGroupInfo (Maybe OutdatedInstancesStrategy)

Indicates what happens when new EC2 instances are launched mid-deployment and do not receive the deployed application revision.

If this option is set to UPDATE or is unspecified, CodeDeploy initiates one or more 'auto-update outdated instances' deployments to apply the deployed application revision to the new EC2 instances.

If this option is set to IGNORE, CodeDeploy does not initiate a deployment to update the new EC2 instances. This may result in instances having different revisions.

deploymentGroupInfo_serviceRoleArn :: Lens' DeploymentGroupInfo (Maybe Text)

A service role Amazon Resource Name (ARN) that grants CodeDeploy permission to make calls to AWS services on your behalf. For more information, see Create a Service Role for AWS CodeDeploy in the AWS CodeDeploy User Guide.

deploymentGroupInfo_deploymentConfigName :: Lens' DeploymentGroupInfo (Maybe Text)

The deployment configuration name.

deploymentGroupInfo_deploymentGroupName :: Lens' DeploymentGroupInfo (Maybe Text)

The deployment group name.

deploymentGroupInfo_autoRollbackConfiguration :: Lens' DeploymentGroupInfo (Maybe AutoRollbackConfiguration)

Information about the automatic rollback configuration associated with the deployment group.

deploymentGroupInfo_deploymentGroupId :: Lens' DeploymentGroupInfo (Maybe Text)

The deployment group ID.

deploymentGroupInfo_triggerConfigurations :: Lens' DeploymentGroupInfo (Maybe [TriggerConfig])

Information about triggers associated with the deployment group.

deploymentGroupInfo_ec2TagFilters :: Lens' DeploymentGroupInfo (Maybe [EC2TagFilter])

The Amazon EC2 tags on which to filter. The deployment group includes EC2 instances with any of the specified tags.

deploymentGroupInfo_targetRevision :: Lens' DeploymentGroupInfo (Maybe RevisionLocation)

Information about the deployment group's target revision, including type and location.

deploymentGroupInfo_lastAttemptedDeployment :: Lens' DeploymentGroupInfo (Maybe LastDeploymentInfo)

Information about the most recent attempted deployment to the deployment group.

deploymentGroupInfo_lastSuccessfulDeployment :: Lens' DeploymentGroupInfo (Maybe LastDeploymentInfo)

Information about the most recent successful deployment to the deployment group.

deploymentGroupInfo_onPremisesInstanceTagFilters :: Lens' DeploymentGroupInfo (Maybe [TagFilter])

The on-premises instance tags on which to filter. The deployment group includes on-premises instances with any of the specified tags.

deploymentGroupInfo_blueGreenDeploymentConfiguration :: Lens' DeploymentGroupInfo (Maybe BlueGreenDeploymentConfiguration)

Information about blue/green deployment options for a deployment group.

deploymentGroupInfo_loadBalancerInfo :: Lens' DeploymentGroupInfo (Maybe LoadBalancerInfo)

Information about the load balancer to use in a deployment.

deploymentGroupInfo_ec2TagSet :: Lens' DeploymentGroupInfo (Maybe EC2TagSet)

Information about groups of tags applied to an EC2 instance. The deployment group includes only EC2 instances identified by all of the tag groups. Cannot be used in the same call as ec2TagFilters.

deploymentGroupInfo_deploymentStyle :: Lens' DeploymentGroupInfo (Maybe DeploymentStyle)

Information about the type of deployment, either in-place or blue/green, you want to run and whether to route deployment traffic behind a load balancer.

deploymentGroupInfo_autoScalingGroups :: Lens' DeploymentGroupInfo (Maybe [AutoScalingGroup])

A list of associated Auto Scaling groups.

deploymentGroupInfo_ecsServices :: Lens' DeploymentGroupInfo (Maybe [ECSService])

The target Amazon ECS services in the deployment group. This applies only to deployment groups that use the Amazon ECS compute platform. A target Amazon ECS service is specified as an Amazon ECS cluster and service name pair using the format <clustername>:<servicename>.

deploymentGroupInfo_alarmConfiguration :: Lens' DeploymentGroupInfo (Maybe AlarmConfiguration)

A list of alarms associated with the deployment group.

deploymentGroupInfo_applicationName :: Lens' DeploymentGroupInfo (Maybe Text)

The application name.

deploymentGroupInfo_computePlatform :: Lens' DeploymentGroupInfo (Maybe ComputePlatform)

The destination platform type for the deployment (Lambda, Server, or ECS).