Network.AWS.CodeDeploy.Types.DeploymentInfo

Description

Documentation

data DeploymentInfo

Information about a deployment.

See: newDeploymentInfo smart constructor.

Constructors

DeploymentInfo'

Fields

deploymentConfigName :: Maybe Text

The deployment configuration name.

deploymentId :: Maybe Text

The unique ID of a deployment.

status :: Maybe DeploymentStatus

The current state of the deployment as a whole.

deploymentGroupName :: Maybe Text

The deployment group name.

autoRollbackConfiguration :: Maybe AutoRollbackConfiguration

Information about the automatic rollback configuration associated with the deployment.

updateOutdatedInstancesOnly :: Maybe Bool

Indicates whether only instances that are not running the latest application revision are to be deployed to.

ignoreApplicationStopFailures :: Maybe Bool

If true, then if an ApplicationStop, BeforeBlockTraffic, or AfterBlockTraffic deployment lifecycle event to an instance fails, then the deployment continues to the next deployment lifecycle event. For example, if ApplicationStop fails, the deployment continues with DownloadBundle. If BeforeBlockTraffic fails, the deployment continues with BlockTraffic. If AfterBlockTraffic fails, the deployment continues with ApplicationStop.

If false or not specified, then if a lifecycle event fails during a deployment to an instance, that deployment fails. If deployment to that instance is part of an overall deployment and the number of healthy hosts is not less than the minimum number of healthy hosts, then a deployment to the next instance is attempted.

During a deployment, the AWS CodeDeploy agent runs the scripts specified for ApplicationStop, BeforeBlockTraffic, and AfterBlockTraffic in the AppSpec file from the previous successful deployment. (All other scripts are run from the AppSpec file in the current deployment.) If one of these scripts contains an error and does not run successfully, the deployment can fail.

If the cause of the failure is a script from the last successful deployment that will never run successfully, create a new deployment and use ignoreApplicationStopFailures to specify that the ApplicationStop, BeforeBlockTraffic, and AfterBlockTraffic failures should be ignored.

targetInstances :: Maybe TargetInstances

Information about the instances that belong to the replacement environment in a blue/green deployment.

startTime :: Maybe POSIX

A timestamp that indicates when the deployment was deployed to the deployment group.

In some cases, the reported value of the start time might be later than the complete time. This is due to differences in the clock settings of backend servers that participate in the deployment process.

instanceTerminationWaitTimeStarted :: Maybe Bool

Indicates whether the wait period set for the termination of instances in the original environment has started. Status is 'false' if the KEEP_ALIVE option is specified. Otherwise, 'true' as soon as the termination wait period starts.

previousRevision :: Maybe RevisionLocation

Information about the application revision that was deployed to the deployment group before the most recent successful deployment.

creator :: Maybe DeploymentCreator

The means by which the deployment was created:

user: A user created the deployment.

autoscaling: Amazon EC2 Auto Scaling created the deployment.

codeDeployRollback: A rollback process created the deployment.

CodeDeployAutoUpdate: An auto-update process created the deployment when it detected outdated EC2 instances.

blueGreenDeploymentConfiguration :: Maybe BlueGreenDeploymentConfiguration

Information about blue/green deployment options for this deployment.

errorInformation :: Maybe ErrorInformation

Information about any error associated with this deployment.

loadBalancerInfo :: Maybe LoadBalancerInfo

Information about the load balancer used in the deployment.

completeTime :: Maybe POSIX

A timestamp that indicates when the deployment was complete.

createTime :: Maybe POSIX

A timestamp that indicates when the deployment was created.

deploymentStyle :: Maybe DeploymentStyle

Information about the type of deployment, either in-place or blue/green, you want to run and whether to route deployment traffic behind a load balancer.

description :: Maybe Text

A comment about the deployment.

revision :: Maybe RevisionLocation

Information about the location of stored application artifacts and the service from which to retrieve them.

rollbackInfo :: Maybe RollbackInfo

Information about a deployment rollback.

externalId :: Maybe Text

The unique ID for an external resource (for example, a CloudFormation stack ID) that is linked to this deployment.

deploymentStatusMessages :: Maybe [Text]

Messages that contain information about the status of a deployment.

relatedDeployments :: Maybe RelatedDeployments

applicationName :: Maybe Text

The application name.

fileExistsBehavior :: Maybe FileExistsBehavior

Information about how AWS CodeDeploy handles files that already exist in a deployment target location but weren't part of the previous successful deployment.

DISALLOW: The deployment fails. This is also the default behavior if no option is specified.

OVERWRITE: The version of the file from the application revision currently being deployed replaces the version already on the instance.

RETAIN: The version of the file already on the instance is kept and used as part of the new deployment.

computePlatform :: Maybe ComputePlatform

The destination platform type for the deployment (Lambda, Server, or ECS).

additionalDeploymentStatusInfo :: Maybe Text

Provides information about the results of a deployment, such as whether instances in the original environment in a blue/green deployment were not terminated.

deploymentOverview :: Maybe DeploymentOverview

A summary of the deployment status of the instances in the deployment.

newDeploymentInfo :: DeploymentInfo

Create a value of DeploymentInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deploymentConfigName:DeploymentInfo', deploymentInfo_deploymentConfigName - The deployment configuration name.

$sel:deploymentId:DeploymentInfo', deploymentInfo_deploymentId - The unique ID of a deployment.

$sel:status:DeploymentInfo', deploymentInfo_status - The current state of the deployment as a whole.

$sel:deploymentGroupName:DeploymentInfo', deploymentInfo_deploymentGroupName - The deployment group name.

$sel:autoRollbackConfiguration:DeploymentInfo', deploymentInfo_autoRollbackConfiguration - Information about the automatic rollback configuration associated with the deployment.

$sel:updateOutdatedInstancesOnly:DeploymentInfo', deploymentInfo_updateOutdatedInstancesOnly - Indicates whether only instances that are not running the latest application revision are to be deployed to.

$sel:ignoreApplicationStopFailures:DeploymentInfo', deploymentInfo_ignoreApplicationStopFailures - If true, then if an ApplicationStop, BeforeBlockTraffic, or AfterBlockTraffic deployment lifecycle event to an instance fails, then the deployment continues to the next deployment lifecycle event. For example, if ApplicationStop fails, the deployment continues with DownloadBundle. If BeforeBlockTraffic fails, the deployment continues with BlockTraffic. If AfterBlockTraffic fails, the deployment continues with ApplicationStop.

If false or not specified, then if a lifecycle event fails during a deployment to an instance, that deployment fails. If deployment to that instance is part of an overall deployment and the number of healthy hosts is not less than the minimum number of healthy hosts, then a deployment to the next instance is attempted.

During a deployment, the AWS CodeDeploy agent runs the scripts specified for ApplicationStop, BeforeBlockTraffic, and AfterBlockTraffic in the AppSpec file from the previous successful deployment. (All other scripts are run from the AppSpec file in the current deployment.) If one of these scripts contains an error and does not run successfully, the deployment can fail.

If the cause of the failure is a script from the last successful deployment that will never run successfully, create a new deployment and use ignoreApplicationStopFailures to specify that the ApplicationStop, BeforeBlockTraffic, and AfterBlockTraffic failures should be ignored.

$sel:targetInstances:DeploymentInfo', deploymentInfo_targetInstances - Information about the instances that belong to the replacement environment in a blue/green deployment.

$sel:startTime:DeploymentInfo', deploymentInfo_startTime - A timestamp that indicates when the deployment was deployed to the deployment group.

In some cases, the reported value of the start time might be later than the complete time. This is due to differences in the clock settings of backend servers that participate in the deployment process.

$sel:instanceTerminationWaitTimeStarted:DeploymentInfo', deploymentInfo_instanceTerminationWaitTimeStarted - Indicates whether the wait period set for the termination of instances in the original environment has started. Status is 'false' if the KEEP_ALIVE option is specified. Otherwise, 'true' as soon as the termination wait period starts.

$sel:previousRevision:DeploymentInfo', deploymentInfo_previousRevision - Information about the application revision that was deployed to the deployment group before the most recent successful deployment.

$sel:creator:DeploymentInfo', deploymentInfo_creator - The means by which the deployment was created:

user: A user created the deployment.

autoscaling: Amazon EC2 Auto Scaling created the deployment.

codeDeployRollback: A rollback process created the deployment.

CodeDeployAutoUpdate: An auto-update process created the deployment when it detected outdated EC2 instances.

$sel:blueGreenDeploymentConfiguration:DeploymentInfo', deploymentInfo_blueGreenDeploymentConfiguration - Information about blue/green deployment options for this deployment.

$sel:errorInformation:DeploymentInfo', deploymentInfo_errorInformation - Information about any error associated with this deployment.

$sel:loadBalancerInfo:DeploymentInfo', deploymentInfo_loadBalancerInfo - Information about the load balancer used in the deployment.

$sel:completeTime:DeploymentInfo', deploymentInfo_completeTime - A timestamp that indicates when the deployment was complete.

$sel:createTime:DeploymentInfo', deploymentInfo_createTime - A timestamp that indicates when the deployment was created.

$sel:deploymentStyle:DeploymentInfo', deploymentInfo_deploymentStyle - Information about the type of deployment, either in-place or blue/green, you want to run and whether to route deployment traffic behind a load balancer.

$sel:description:DeploymentInfo', deploymentInfo_description - A comment about the deployment.

$sel:revision:DeploymentInfo', deploymentInfo_revision - Information about the location of stored application artifacts and the service from which to retrieve them.

$sel:rollbackInfo:DeploymentInfo', deploymentInfo_rollbackInfo - Information about a deployment rollback.

$sel:externalId:DeploymentInfo', deploymentInfo_externalId - The unique ID for an external resource (for example, a CloudFormation stack ID) that is linked to this deployment.

$sel:deploymentStatusMessages:DeploymentInfo', deploymentInfo_deploymentStatusMessages - Messages that contain information about the status of a deployment.

$sel:relatedDeployments:DeploymentInfo', deploymentInfo_relatedDeployments - Undocumented member.

$sel:applicationName:DeploymentInfo', deploymentInfo_applicationName - The application name.

$sel:fileExistsBehavior:DeploymentInfo', deploymentInfo_fileExistsBehavior - Information about how AWS CodeDeploy handles files that already exist in a deployment target location but weren't part of the previous successful deployment.

DISALLOW: The deployment fails. This is also the default behavior if no option is specified.

OVERWRITE: The version of the file from the application revision currently being deployed replaces the version already on the instance.

RETAIN: The version of the file already on the instance is kept and used as part of the new deployment.

$sel:computePlatform:DeploymentInfo', deploymentInfo_computePlatform - The destination platform type for the deployment (Lambda, Server, or ECS).

$sel:additionalDeploymentStatusInfo:DeploymentInfo', deploymentInfo_additionalDeploymentStatusInfo - Provides information about the results of a deployment, such as whether instances in the original environment in a blue/green deployment were not terminated.

$sel:deploymentOverview:DeploymentInfo', deploymentInfo_deploymentOverview - A summary of the deployment status of the instances in the deployment.

deploymentInfo_deploymentConfigName :: Lens' DeploymentInfo (Maybe Text)

The deployment configuration name.

deploymentInfo_deploymentId :: Lens' DeploymentInfo (Maybe Text)

The unique ID of a deployment.

deploymentInfo_status :: Lens' DeploymentInfo (Maybe DeploymentStatus)

The current state of the deployment as a whole.

deploymentInfo_deploymentGroupName :: Lens' DeploymentInfo (Maybe Text)

The deployment group name.

deploymentInfo_autoRollbackConfiguration :: Lens' DeploymentInfo (Maybe AutoRollbackConfiguration)

Information about the automatic rollback configuration associated with the deployment.

deploymentInfo_updateOutdatedInstancesOnly :: Lens' DeploymentInfo (Maybe Bool)

Indicates whether only instances that are not running the latest application revision are to be deployed to.

deploymentInfo_ignoreApplicationStopFailures :: Lens' DeploymentInfo (Maybe Bool)

If true, then if an ApplicationStop, BeforeBlockTraffic, or AfterBlockTraffic deployment lifecycle event to an instance fails, then the deployment continues to the next deployment lifecycle event. For example, if ApplicationStop fails, the deployment continues with DownloadBundle. If BeforeBlockTraffic fails, the deployment continues with BlockTraffic. If AfterBlockTraffic fails, the deployment continues with ApplicationStop.

If false or not specified, then if a lifecycle event fails during a deployment to an instance, that deployment fails. If deployment to that instance is part of an overall deployment and the number of healthy hosts is not less than the minimum number of healthy hosts, then a deployment to the next instance is attempted.

During a deployment, the AWS CodeDeploy agent runs the scripts specified for ApplicationStop, BeforeBlockTraffic, and AfterBlockTraffic in the AppSpec file from the previous successful deployment. (All other scripts are run from the AppSpec file in the current deployment.) If one of these scripts contains an error and does not run successfully, the deployment can fail.

If the cause of the failure is a script from the last successful deployment that will never run successfully, create a new deployment and use ignoreApplicationStopFailures to specify that the ApplicationStop, BeforeBlockTraffic, and AfterBlockTraffic failures should be ignored.

deploymentInfo_targetInstances :: Lens' DeploymentInfo (Maybe TargetInstances)

Information about the instances that belong to the replacement environment in a blue/green deployment.

deploymentInfo_startTime :: Lens' DeploymentInfo (Maybe UTCTime)

A timestamp that indicates when the deployment was deployed to the deployment group.

In some cases, the reported value of the start time might be later than the complete time. This is due to differences in the clock settings of backend servers that participate in the deployment process.

deploymentInfo_instanceTerminationWaitTimeStarted :: Lens' DeploymentInfo (Maybe Bool)

Indicates whether the wait period set for the termination of instances in the original environment has started. Status is 'false' if the KEEP_ALIVE option is specified. Otherwise, 'true' as soon as the termination wait period starts.

deploymentInfo_previousRevision :: Lens' DeploymentInfo (Maybe RevisionLocation)

Information about the application revision that was deployed to the deployment group before the most recent successful deployment.

deploymentInfo_creator :: Lens' DeploymentInfo (Maybe DeploymentCreator)

The means by which the deployment was created:

user: A user created the deployment.

autoscaling: Amazon EC2 Auto Scaling created the deployment.

codeDeployRollback: A rollback process created the deployment.

CodeDeployAutoUpdate: An auto-update process created the deployment when it detected outdated EC2 instances.

deploymentInfo_blueGreenDeploymentConfiguration :: Lens' DeploymentInfo (Maybe BlueGreenDeploymentConfiguration)

Information about blue/green deployment options for this deployment.

deploymentInfo_errorInformation :: Lens' DeploymentInfo (Maybe ErrorInformation)

Information about any error associated with this deployment.

deploymentInfo_loadBalancerInfo :: Lens' DeploymentInfo (Maybe LoadBalancerInfo)

Information about the load balancer used in the deployment.

deploymentInfo_completeTime :: Lens' DeploymentInfo (Maybe UTCTime)

A timestamp that indicates when the deployment was complete.

deploymentInfo_createTime :: Lens' DeploymentInfo (Maybe UTCTime)

A timestamp that indicates when the deployment was created.

deploymentInfo_deploymentStyle :: Lens' DeploymentInfo (Maybe DeploymentStyle)

Information about the type of deployment, either in-place or blue/green, you want to run and whether to route deployment traffic behind a load balancer.

deploymentInfo_description :: Lens' DeploymentInfo (Maybe Text)

A comment about the deployment.

deploymentInfo_revision :: Lens' DeploymentInfo (Maybe RevisionLocation)

Information about the location of stored application artifacts and the service from which to retrieve them.

deploymentInfo_rollbackInfo :: Lens' DeploymentInfo (Maybe RollbackInfo)

Information about a deployment rollback.

deploymentInfo_externalId :: Lens' DeploymentInfo (Maybe Text)

The unique ID for an external resource (for example, a CloudFormation stack ID) that is linked to this deployment.

deploymentInfo_deploymentStatusMessages :: Lens' DeploymentInfo (Maybe [Text])

Messages that contain information about the status of a deployment.

deploymentInfo_relatedDeployments :: Lens' DeploymentInfo (Maybe RelatedDeployments)

Undocumented member.

deploymentInfo_applicationName :: Lens' DeploymentInfo (Maybe Text)

The application name.

deploymentInfo_fileExistsBehavior :: Lens' DeploymentInfo (Maybe FileExistsBehavior)

Information about how AWS CodeDeploy handles files that already exist in a deployment target location but weren't part of the previous successful deployment.

DISALLOW: The deployment fails. This is also the default behavior if no option is specified.

OVERWRITE: The version of the file from the application revision currently being deployed replaces the version already on the instance.

RETAIN: The version of the file already on the instance is kept and used as part of the new deployment.

deploymentInfo_computePlatform :: Lens' DeploymentInfo (Maybe ComputePlatform)

The destination platform type for the deployment (Lambda, Server, or ECS).

deploymentInfo_additionalDeploymentStatusInfo :: Lens' DeploymentInfo (Maybe Text)

Provides information about the results of a deployment, such as whether instances in the original environment in a blue/green deployment were not terminated.

deploymentInfo_deploymentOverview :: Lens' DeploymentInfo (Maybe DeploymentOverview)

A summary of the deployment status of the instances in the deployment.