Network.AWS.CodeDeploy.Types.DeploymentOverview

Description

Documentation

data DeploymentOverview

Information about the deployment status of the instances in the deployment.

See: newDeploymentOverview smart constructor.

Constructors

DeploymentOverview'

Fields

pending :: Maybe Integer

The number of instances in the deployment in a pending state.

skipped :: Maybe Integer

The number of instances in the deployment in a skipped state.

inProgress :: Maybe Integer

The number of instances in which the deployment is in progress.

succeeded :: Maybe Integer

The number of instances in the deployment to which revisions have been successfully deployed.

ready :: Maybe Integer

The number of instances in a replacement environment ready to receive traffic in a blue/green deployment.

failed :: Maybe Integer

The number of instances in the deployment in a failed state.

newDeploymentOverview :: DeploymentOverview

Create a value of DeploymentOverview with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pending:DeploymentOverview', deploymentOverview_pending - The number of instances in the deployment in a pending state.

$sel:skipped:DeploymentOverview', deploymentOverview_skipped - The number of instances in the deployment in a skipped state.

$sel:inProgress:DeploymentOverview', deploymentOverview_inProgress - The number of instances in which the deployment is in progress.

$sel:succeeded:DeploymentOverview', deploymentOverview_succeeded - The number of instances in the deployment to which revisions have been successfully deployed.

$sel:ready:DeploymentOverview', deploymentOverview_ready - The number of instances in a replacement environment ready to receive traffic in a blue/green deployment.

$sel:failed:DeploymentOverview', deploymentOverview_failed - The number of instances in the deployment in a failed state.

deploymentOverview_pending :: Lens' DeploymentOverview (Maybe Integer)

The number of instances in the deployment in a pending state.

deploymentOverview_skipped :: Lens' DeploymentOverview (Maybe Integer)

The number of instances in the deployment in a skipped state.

deploymentOverview_inProgress :: Lens' DeploymentOverview (Maybe Integer)

The number of instances in which the deployment is in progress.

deploymentOverview_succeeded :: Lens' DeploymentOverview (Maybe Integer)

The number of instances in the deployment to which revisions have been successfully deployed.

deploymentOverview_ready :: Lens' DeploymentOverview (Maybe Integer)

The number of instances in a replacement environment ready to receive traffic in a blue/green deployment.

deploymentOverview_failed :: Lens' DeploymentOverview (Maybe Integer)

The number of instances in the deployment in a failed state.