Network.AWS.CodeDeploy.Types.DeploymentTarget

Description

Documentation

data DeploymentTarget

Information about the deployment target.

See: newDeploymentTarget smart constructor.

Constructors

DeploymentTarget'

Fields

ecsTarget :: Maybe ECSTarget

Information about the target for a deployment that uses the Amazon ECS compute platform.

lambdaTarget :: Maybe LambdaTarget

Information about the target for a deployment that uses the AWS Lambda compute platform.

cloudFormationTarget :: Maybe CloudFormationTarget

instanceTarget :: Maybe InstanceTarget

Information about the target for a deployment that uses the EC2/On-premises compute platform.

deploymentTargetType :: Maybe DeploymentTargetType

The deployment type that is specific to the deployment's compute platform or deployments initiated by a CloudFormation stack update.

newDeploymentTarget :: DeploymentTarget

Create a value of DeploymentTarget with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ecsTarget:DeploymentTarget', deploymentTarget_ecsTarget - Information about the target for a deployment that uses the Amazon ECS compute platform.

$sel:lambdaTarget:DeploymentTarget', deploymentTarget_lambdaTarget - Information about the target for a deployment that uses the AWS Lambda compute platform.

$sel:cloudFormationTarget:DeploymentTarget', deploymentTarget_cloudFormationTarget - Undocumented member.

$sel:instanceTarget:DeploymentTarget', deploymentTarget_instanceTarget - Information about the target for a deployment that uses the EC2/On-premises compute platform.

$sel:deploymentTargetType:DeploymentTarget', deploymentTarget_deploymentTargetType - The deployment type that is specific to the deployment's compute platform or deployments initiated by a CloudFormation stack update.

deploymentTarget_ecsTarget :: Lens' DeploymentTarget (Maybe ECSTarget)

Information about the target for a deployment that uses the Amazon ECS compute platform.

deploymentTarget_lambdaTarget :: Lens' DeploymentTarget (Maybe LambdaTarget)

Information about the target for a deployment that uses the AWS Lambda compute platform.

deploymentTarget_cloudFormationTarget :: Lens' DeploymentTarget (Maybe CloudFormationTarget)

Undocumented member.

deploymentTarget_instanceTarget :: Lens' DeploymentTarget (Maybe InstanceTarget)

Information about the target for a deployment that uses the EC2/On-premises compute platform.

deploymentTarget_deploymentTargetType :: Lens' DeploymentTarget (Maybe DeploymentTargetType)

The deployment type that is specific to the deployment's compute platform or deployments initiated by a CloudFormation stack update.