Network.AWS.CodeDeploy.Types.ECSTaskSet

Description

Documentation

data ECSTaskSet

Information about a set of Amazon ECS tasks in an AWS CodeDeploy deployment. An Amazon ECS task set includes details such as the desired number of tasks, how many tasks are running, and whether the task set serves production traffic. An AWS CodeDeploy application that uses the Amazon ECS compute platform deploys a containerized application in an Amazon ECS service as a task set.

See: newECSTaskSet smart constructor.

Constructors

ECSTaskSet'

Fields

runningCount :: Maybe Integer

The number of tasks in the task set that are in the RUNNING status during an Amazon ECS deployment. A task in the RUNNING state is running and ready for use.

status :: Maybe Text

The status of the task set. There are three valid task set statuses:

PRIMARY: Indicates the task set is serving production traffic.

ACTIVE: Indicates the task set is not serving production traffic.

DRAINING: Indicates the tasks in the task set are being stopped and their corresponding targets are being deregistered from their target group.

desiredCount :: Maybe Integer

The number of tasks in a task set. During a deployment that uses the Amazon ECS compute type, CodeDeploy instructs Amazon ECS to create a new task set and uses this value to determine how many tasks to create. After the updated task set is created, CodeDeploy shifts traffic to the new task set.

pendingCount :: Maybe Integer

The number of tasks in the task set that are in the PENDING status during an Amazon ECS deployment. A task in the PENDING state is preparing to enter the RUNNING state. A task set enters the PENDING status when it launches for the first time, or when it is restarted after being in the STOPPED state.

taskSetLabel :: Maybe TargetLabel

A label that identifies whether the ECS task set is an original target (BLUE) or a replacement target (GREEN).

targetGroup :: Maybe TargetGroupInfo

The target group associated with the task set. The target group is used by AWS CodeDeploy to manage traffic to a task set.

trafficWeight :: Maybe Double

The percentage of traffic served by this task set.

identifer :: Maybe Text

A unique ID of an ECSTaskSet.

newECSTaskSet :: ECSTaskSet

Create a value of ECSTaskSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:runningCount:ECSTaskSet', eCSTaskSet_runningCount - The number of tasks in the task set that are in the RUNNING status during an Amazon ECS deployment. A task in the RUNNING state is running and ready for use.

$sel:status:ECSTaskSet', eCSTaskSet_status - The status of the task set. There are three valid task set statuses:

PRIMARY: Indicates the task set is serving production traffic.

ACTIVE: Indicates the task set is not serving production traffic.

DRAINING: Indicates the tasks in the task set are being stopped and their corresponding targets are being deregistered from their target group.

$sel:desiredCount:ECSTaskSet', eCSTaskSet_desiredCount - The number of tasks in a task set. During a deployment that uses the Amazon ECS compute type, CodeDeploy instructs Amazon ECS to create a new task set and uses this value to determine how many tasks to create. After the updated task set is created, CodeDeploy shifts traffic to the new task set.

$sel:pendingCount:ECSTaskSet', eCSTaskSet_pendingCount - The number of tasks in the task set that are in the PENDING status during an Amazon ECS deployment. A task in the PENDING state is preparing to enter the RUNNING state. A task set enters the PENDING status when it launches for the first time, or when it is restarted after being in the STOPPED state.

$sel:taskSetLabel:ECSTaskSet', eCSTaskSet_taskSetLabel - A label that identifies whether the ECS task set is an original target (BLUE) or a replacement target (GREEN).

$sel:targetGroup:ECSTaskSet', eCSTaskSet_targetGroup - The target group associated with the task set. The target group is used by AWS CodeDeploy to manage traffic to a task set.

$sel:trafficWeight:ECSTaskSet', eCSTaskSet_trafficWeight - The percentage of traffic served by this task set.

$sel:identifer:ECSTaskSet', eCSTaskSet_identifer - A unique ID of an ECSTaskSet.

eCSTaskSet_runningCount :: Lens' ECSTaskSet (Maybe Integer)

The number of tasks in the task set that are in the RUNNING status during an Amazon ECS deployment. A task in the RUNNING state is running and ready for use.

eCSTaskSet_status :: Lens' ECSTaskSet (Maybe Text)

The status of the task set. There are three valid task set statuses:

PRIMARY: Indicates the task set is serving production traffic.

ACTIVE: Indicates the task set is not serving production traffic.

DRAINING: Indicates the tasks in the task set are being stopped and their corresponding targets are being deregistered from their target group.

eCSTaskSet_desiredCount :: Lens' ECSTaskSet (Maybe Integer)

The number of tasks in a task set. During a deployment that uses the Amazon ECS compute type, CodeDeploy instructs Amazon ECS to create a new task set and uses this value to determine how many tasks to create. After the updated task set is created, CodeDeploy shifts traffic to the new task set.

eCSTaskSet_pendingCount :: Lens' ECSTaskSet (Maybe Integer)

The number of tasks in the task set that are in the PENDING status during an Amazon ECS deployment. A task in the PENDING state is preparing to enter the RUNNING state. A task set enters the PENDING status when it launches for the first time, or when it is restarted after being in the STOPPED state.

eCSTaskSet_taskSetLabel :: Lens' ECSTaskSet (Maybe TargetLabel)

A label that identifies whether the ECS task set is an original target (BLUE) or a replacement target (GREEN).

eCSTaskSet_targetGroup :: Lens' ECSTaskSet (Maybe TargetGroupInfo)

The target group associated with the task set. The target group is used by AWS CodeDeploy to manage traffic to a task set.

eCSTaskSet_trafficWeight :: Lens' ECSTaskSet (Maybe Double)

The percentage of traffic served by this task set.

eCSTaskSet_identifer :: Lens' ECSTaskSet (Maybe Text)

A unique ID of an ECSTaskSet.