Network.AWS.CodeDeploy.Types.GitHubLocation

Description

Documentation

data GitHubLocation

Information about the location of application artifacts stored in GitHub.

See: newGitHubLocation smart constructor.

Constructors

GitHubLocation'

Fields

commitId :: Maybe Text

The SHA1 commit ID of the GitHub commit that represents the bundled artifacts for the application revision.

repository :: Maybe Text

The GitHub account and repository pair that stores a reference to the commit that represents the bundled artifacts for the application revision.

Specified as account/repository.

newGitHubLocation :: GitHubLocation

Create a value of GitHubLocation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:commitId:GitHubLocation', gitHubLocation_commitId - The SHA1 commit ID of the GitHub commit that represents the bundled artifacts for the application revision.

$sel:repository:GitHubLocation', gitHubLocation_repository - The GitHub account and repository pair that stores a reference to the commit that represents the bundled artifacts for the application revision.

Specified as account/repository.

gitHubLocation_commitId :: Lens' GitHubLocation (Maybe Text)

The SHA1 commit ID of the GitHub commit that represents the bundled artifacts for the application revision.

gitHubLocation_repository :: Lens' GitHubLocation (Maybe Text)

The GitHub account and repository pair that stores a reference to the commit that represents the bundled artifacts for the application revision.

Specified as account/repository.