Network.AWS.CodeDeploy.Types.GreenFleetProvisioningOption

Description

Documentation

data GreenFleetProvisioningOption

Information about the instances that belong to the replacement environment in a blue/green deployment.

See: newGreenFleetProvisioningOption smart constructor.

Constructors

GreenFleetProvisioningOption'

Fields

action :: Maybe GreenFleetProvisioningAction

The method used to add instances to a replacement environment.

DISCOVER_EXISTING: Use instances that already exist or will be created manually.

COPY_AUTO_SCALING_GROUP: Use settings from a specified Auto Scaling group to define and create instances in a new Auto Scaling group.

newGreenFleetProvisioningOption :: GreenFleetProvisioningOption

Create a value of GreenFleetProvisioningOption with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:action:GreenFleetProvisioningOption', greenFleetProvisioningOption_action - The method used to add instances to a replacement environment.

DISCOVER_EXISTING: Use instances that already exist or will be created manually.

COPY_AUTO_SCALING_GROUP: Use settings from a specified Auto Scaling group to define and create instances in a new Auto Scaling group.

greenFleetProvisioningOption_action :: Lens' GreenFleetProvisioningOption (Maybe GreenFleetProvisioningAction)

The method used to add instances to a replacement environment.

DISCOVER_EXISTING: Use instances that already exist or will be created manually.

COPY_AUTO_SCALING_GROUP: Use settings from a specified Auto Scaling group to define and create instances in a new Auto Scaling group.