Network.AWS.CodeDeploy.Types.InstanceInfo

Description

Documentation

data InstanceInfo

Information about an on-premises instance.

See: newInstanceInfo smart constructor.

Constructors

InstanceInfo'

Fields

registerTime :: Maybe POSIX

The time at which the on-premises instance was registered.

instanceArn :: Maybe Text

The ARN of the on-premises instance.

deregisterTime :: Maybe POSIX

If the on-premises instance was deregistered, the time at which the on-premises instance was deregistered.

iamUserArn :: Maybe Text

The IAM user ARN associated with the on-premises instance.

instanceName :: Maybe Text

The name of the on-premises instance.

iamSessionArn :: Maybe Text

The ARN of the IAM session associated with the on-premises instance.

tags :: Maybe [Tag]

The tags currently associated with the on-premises instance.

newInstanceInfo :: InstanceInfo

Create a value of InstanceInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:registerTime:InstanceInfo', instanceInfo_registerTime - The time at which the on-premises instance was registered.

$sel:instanceArn:InstanceInfo', instanceInfo_instanceArn - The ARN of the on-premises instance.

$sel:deregisterTime:InstanceInfo', instanceInfo_deregisterTime - If the on-premises instance was deregistered, the time at which the on-premises instance was deregistered.

$sel:iamUserArn:InstanceInfo', instanceInfo_iamUserArn - The IAM user ARN associated with the on-premises instance.

$sel:instanceName:InstanceInfo', instanceInfo_instanceName - The name of the on-premises instance.

$sel:iamSessionArn:InstanceInfo', instanceInfo_iamSessionArn - The ARN of the IAM session associated with the on-premises instance.

$sel:tags:InstanceInfo', instanceInfo_tags - The tags currently associated with the on-premises instance.

instanceInfo_registerTime :: Lens' InstanceInfo (Maybe UTCTime)

The time at which the on-premises instance was registered.

instanceInfo_instanceArn :: Lens' InstanceInfo (Maybe Text)

The ARN of the on-premises instance.

instanceInfo_deregisterTime :: Lens' InstanceInfo (Maybe UTCTime)

If the on-premises instance was deregistered, the time at which the on-premises instance was deregistered.

instanceInfo_iamUserArn :: Lens' InstanceInfo (Maybe Text)

The IAM user ARN associated with the on-premises instance.

instanceInfo_instanceName :: Lens' InstanceInfo (Maybe Text)

The name of the on-premises instance.

instanceInfo_iamSessionArn :: Lens' InstanceInfo (Maybe Text)

The ARN of the IAM session associated with the on-premises instance.

instanceInfo_tags :: Lens' InstanceInfo (Maybe [Tag])

The tags currently associated with the on-premises instance.