Network.AWS.CodeDeploy.Types.InstanceTarget

Description

Documentation

data InstanceTarget

A target Amazon EC2 or on-premises instance during a deployment that uses the EC2/On-premises compute platform.

See: newInstanceTarget smart constructor.

Constructors

InstanceTarget'

Fields

targetId :: Maybe Text

The unique ID of a deployment target that has a type of instanceTarget.

deploymentId :: Maybe Text

The unique ID of a deployment.

status :: Maybe TargetStatus

The status an EC2/On-premises deployment's target instance.

instanceLabel :: Maybe TargetLabel

A label that identifies whether the instance is an original target (BLUE) or a replacement target (GREEN).

targetArn :: Maybe Text

The Amazon Resource Name (ARN) of the target.

lifecycleEvents :: Maybe [LifecycleEvent]

The lifecycle events of the deployment to this target instance.

lastUpdatedAt :: Maybe POSIX

The date and time when the target instance was updated by a deployment.

newInstanceTarget :: InstanceTarget

Create a value of InstanceTarget with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetId:InstanceTarget', instanceTarget_targetId - The unique ID of a deployment target that has a type of instanceTarget.

$sel:deploymentId:InstanceTarget', instanceTarget_deploymentId - The unique ID of a deployment.

$sel:status:InstanceTarget', instanceTarget_status - The status an EC2/On-premises deployment's target instance.

$sel:instanceLabel:InstanceTarget', instanceTarget_instanceLabel - A label that identifies whether the instance is an original target (BLUE) or a replacement target (GREEN).

$sel:targetArn:InstanceTarget', instanceTarget_targetArn - The Amazon Resource Name (ARN) of the target.

$sel:lifecycleEvents:InstanceTarget', instanceTarget_lifecycleEvents - The lifecycle events of the deployment to this target instance.

$sel:lastUpdatedAt:InstanceTarget', instanceTarget_lastUpdatedAt - The date and time when the target instance was updated by a deployment.

instanceTarget_targetId :: Lens' InstanceTarget (Maybe Text)

The unique ID of a deployment target that has a type of instanceTarget.

instanceTarget_deploymentId :: Lens' InstanceTarget (Maybe Text)

The unique ID of a deployment.

instanceTarget_status :: Lens' InstanceTarget (Maybe TargetStatus)

The status an EC2/On-premises deployment's target instance.

instanceTarget_instanceLabel :: Lens' InstanceTarget (Maybe TargetLabel)

A label that identifies whether the instance is an original target (BLUE) or a replacement target (GREEN).

instanceTarget_targetArn :: Lens' InstanceTarget (Maybe Text)

The Amazon Resource Name (ARN) of the target.

instanceTarget_lifecycleEvents :: Lens' InstanceTarget (Maybe [LifecycleEvent])

The lifecycle events of the deployment to this target instance.

instanceTarget_lastUpdatedAt :: Lens' InstanceTarget (Maybe UTCTime)

The date and time when the target instance was updated by a deployment.