Network.AWS.CodeDeploy.Types.LambdaTarget

Description

Documentation

data LambdaTarget

Information about the target AWS Lambda function during an AWS Lambda deployment.

See: newLambdaTarget smart constructor.

Constructors

LambdaTarget'

Fields

targetId :: Maybe Text

The unique ID of a deployment target that has a type of lambdaTarget.

deploymentId :: Maybe Text

The unique ID of a deployment.

status :: Maybe TargetStatus

The status an AWS Lambda deployment's target Lambda function.

targetArn :: Maybe Text

The Amazon Resource Name (ARN) of the target.

lifecycleEvents :: Maybe [LifecycleEvent]

The lifecycle events of the deployment to this target Lambda function.

lambdaFunctionInfo :: Maybe LambdaFunctionInfo

A LambdaFunctionInfo object that describes a target Lambda function.

lastUpdatedAt :: Maybe POSIX

The date and time when the target Lambda function was updated by a deployment.

newLambdaTarget :: LambdaTarget

Create a value of LambdaTarget with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetId:LambdaTarget', lambdaTarget_targetId - The unique ID of a deployment target that has a type of lambdaTarget.

$sel:deploymentId:LambdaTarget', lambdaTarget_deploymentId - The unique ID of a deployment.

$sel:status:LambdaTarget', lambdaTarget_status - The status an AWS Lambda deployment's target Lambda function.

$sel:targetArn:LambdaTarget', lambdaTarget_targetArn - The Amazon Resource Name (ARN) of the target.

$sel:lifecycleEvents:LambdaTarget', lambdaTarget_lifecycleEvents - The lifecycle events of the deployment to this target Lambda function.

$sel:lambdaFunctionInfo:LambdaTarget', lambdaTarget_lambdaFunctionInfo - A LambdaFunctionInfo object that describes a target Lambda function.

$sel:lastUpdatedAt:LambdaTarget', lambdaTarget_lastUpdatedAt - The date and time when the target Lambda function was updated by a deployment.

lambdaTarget_targetId :: Lens' LambdaTarget (Maybe Text)

The unique ID of a deployment target that has a type of lambdaTarget.

lambdaTarget_deploymentId :: Lens' LambdaTarget (Maybe Text)

The unique ID of a deployment.

lambdaTarget_status :: Lens' LambdaTarget (Maybe TargetStatus)

The status an AWS Lambda deployment's target Lambda function.

lambdaTarget_targetArn :: Lens' LambdaTarget (Maybe Text)

The Amazon Resource Name (ARN) of the target.

lambdaTarget_lifecycleEvents :: Lens' LambdaTarget (Maybe [LifecycleEvent])

The lifecycle events of the deployment to this target Lambda function.

lambdaTarget_lambdaFunctionInfo :: Lens' LambdaTarget (Maybe LambdaFunctionInfo)

A LambdaFunctionInfo object that describes a target Lambda function.

lambdaTarget_lastUpdatedAt :: Lens' LambdaTarget (Maybe UTCTime)

The date and time when the target Lambda function was updated by a deployment.