Network.AWS.CodeDeploy.Types.LifecycleEvent

Description

Documentation

data LifecycleEvent

Information about a deployment lifecycle event.

See: newLifecycleEvent smart constructor.

Constructors

LifecycleEvent'

Fields

diagnostics :: Maybe Diagnostics

Diagnostic information about the deployment lifecycle event.

status :: Maybe LifecycleEventStatus

The deployment lifecycle event status:

Pending: The deployment lifecycle event is pending.

InProgress: The deployment lifecycle event is in progress.

Succeeded: The deployment lifecycle event ran successfully.

Failed: The deployment lifecycle event has failed.

Skipped: The deployment lifecycle event has been skipped.

Unknown: The deployment lifecycle event is unknown.

startTime :: Maybe POSIX

A timestamp that indicates when the deployment lifecycle event started.

endTime :: Maybe POSIX

A timestamp that indicates when the deployment lifecycle event ended.

lifecycleEventName :: Maybe Text

The deployment lifecycle event name, such as ApplicationStop, BeforeInstall, AfterInstall, ApplicationStart, or ValidateService.

newLifecycleEvent :: LifecycleEvent

Create a value of LifecycleEvent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:diagnostics:LifecycleEvent', lifecycleEvent_diagnostics - Diagnostic information about the deployment lifecycle event.

$sel:status:LifecycleEvent', lifecycleEvent_status - The deployment lifecycle event status:

Pending: The deployment lifecycle event is pending.

InProgress: The deployment lifecycle event is in progress.

Succeeded: The deployment lifecycle event ran successfully.

Failed: The deployment lifecycle event has failed.

Skipped: The deployment lifecycle event has been skipped.

Unknown: The deployment lifecycle event is unknown.

$sel:startTime:LifecycleEvent', lifecycleEvent_startTime - A timestamp that indicates when the deployment lifecycle event started.

$sel:endTime:LifecycleEvent', lifecycleEvent_endTime - A timestamp that indicates when the deployment lifecycle event ended.

$sel:lifecycleEventName:LifecycleEvent', lifecycleEvent_lifecycleEventName - The deployment lifecycle event name, such as ApplicationStop, BeforeInstall, AfterInstall, ApplicationStart, or ValidateService.

lifecycleEvent_diagnostics :: Lens' LifecycleEvent (Maybe Diagnostics)

Diagnostic information about the deployment lifecycle event.

lifecycleEvent_status :: Lens' LifecycleEvent (Maybe LifecycleEventStatus)

The deployment lifecycle event status:

Pending: The deployment lifecycle event is pending.

InProgress: The deployment lifecycle event is in progress.

Succeeded: The deployment lifecycle event ran successfully.

Failed: The deployment lifecycle event has failed.

Skipped: The deployment lifecycle event has been skipped.

Unknown: The deployment lifecycle event is unknown.

lifecycleEvent_startTime :: Lens' LifecycleEvent (Maybe UTCTime)

A timestamp that indicates when the deployment lifecycle event started.

lifecycleEvent_endTime :: Lens' LifecycleEvent (Maybe UTCTime)

A timestamp that indicates when the deployment lifecycle event ended.

lifecycleEvent_lifecycleEventName :: Lens' LifecycleEvent (Maybe Text)

The deployment lifecycle event name, such as ApplicationStop, BeforeInstall, AfterInstall, ApplicationStart, or ValidateService.