Network.AWS.CodeDeploy.Types.RevisionLocation

Description

Documentation

data RevisionLocation

Information about the location of an application revision.

See: newRevisionLocation smart constructor.

Constructors

RevisionLocation'

Fields

string :: Maybe RawString

Information about the location of an AWS Lambda deployment revision stored as a RawString.

revisionType :: Maybe RevisionLocationType

The type of application revision:

S3: An application revision stored in Amazon S3.

GitHub: An application revision stored in GitHub (EC2/On-premises deployments only).

String: A YAML-formatted or JSON-formatted string (AWS Lambda deployments only).

AppSpecContent: An AppSpecContent object that contains the contents of an AppSpec file for an AWS Lambda or Amazon ECS deployment. The content is formatted as JSON or YAML stored as a RawString.

s3Location :: Maybe S3Location

Information about the location of a revision stored in Amazon S3.

appSpecContent :: Maybe AppSpecContent

The content of an AppSpec file for an AWS Lambda or Amazon ECS deployment. The content is formatted as JSON or YAML and stored as a RawString.

gitHubLocation :: Maybe GitHubLocation

Information about the location of application artifacts stored in GitHub.

newRevisionLocation :: RevisionLocation

Create a value of RevisionLocation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:string:RevisionLocation', revisionLocation_string - Information about the location of an AWS Lambda deployment revision stored as a RawString.

$sel:revisionType:RevisionLocation', revisionLocation_revisionType - The type of application revision:

S3: An application revision stored in Amazon S3.

GitHub: An application revision stored in GitHub (EC2/On-premises deployments only).

String: A YAML-formatted or JSON-formatted string (AWS Lambda deployments only).

AppSpecContent: An AppSpecContent object that contains the contents of an AppSpec file for an AWS Lambda or Amazon ECS deployment. The content is formatted as JSON or YAML stored as a RawString.

$sel:s3Location:RevisionLocation', revisionLocation_s3Location - Information about the location of a revision stored in Amazon S3.

$sel:appSpecContent:RevisionLocation', revisionLocation_appSpecContent - The content of an AppSpec file for an AWS Lambda or Amazon ECS deployment. The content is formatted as JSON or YAML and stored as a RawString.

$sel:gitHubLocation:RevisionLocation', revisionLocation_gitHubLocation - Information about the location of application artifacts stored in GitHub.

revisionLocation_string :: Lens' RevisionLocation (Maybe RawString)

Information about the location of an AWS Lambda deployment revision stored as a RawString.

revisionLocation_revisionType :: Lens' RevisionLocation (Maybe RevisionLocationType)

The type of application revision:

S3: An application revision stored in Amazon S3.

GitHub: An application revision stored in GitHub (EC2/On-premises deployments only).

String: A YAML-formatted or JSON-formatted string (AWS Lambda deployments only).

AppSpecContent: An AppSpecContent object that contains the contents of an AppSpec file for an AWS Lambda or Amazon ECS deployment. The content is formatted as JSON or YAML stored as a RawString.

revisionLocation_s3Location :: Lens' RevisionLocation (Maybe S3Location)

Information about the location of a revision stored in Amazon S3.

revisionLocation_appSpecContent :: Lens' RevisionLocation (Maybe AppSpecContent)

The content of an AppSpec file for an AWS Lambda or Amazon ECS deployment. The content is formatted as JSON or YAML and stored as a RawString.

revisionLocation_gitHubLocation :: Lens' RevisionLocation (Maybe GitHubLocation)

Information about the location of application artifacts stored in GitHub.