Network.AWS.CodeDeploy.Types.S3Location

Description

Documentation

data S3Location

Information about the location of application artifacts stored in Amazon S3.

See: newS3Location smart constructor.

Constructors

S3Location'

Fields

bundleType :: Maybe BundleType

The file type of the application revision. Must be one of the following:

tar: A tar archive file.

tgz: A compressed tar archive file.

zip: A zip archive file.

eTag :: Maybe Text

The ETag of the Amazon S3 object that represents the bundled artifacts for the application revision.

If the ETag is not specified as an input parameter, ETag validation of the object is skipped.

bucket :: Maybe Text

The name of the Amazon S3 bucket where the application revision is stored.

key :: Maybe Text

The name of the Amazon S3 object that represents the bundled artifacts for the application revision.

version :: Maybe Text

A specific version of the Amazon S3 object that represents the bundled artifacts for the application revision.

If the version is not specified, the system uses the most recent version by default.

newS3Location :: S3Location

Create a value of S3Location with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bundleType:S3Location', s3Location_bundleType - The file type of the application revision. Must be one of the following:

tar: A tar archive file.

tgz: A compressed tar archive file.

zip: A zip archive file.

$sel:eTag:S3Location', s3Location_eTag - The ETag of the Amazon S3 object that represents the bundled artifacts for the application revision.

If the ETag is not specified as an input parameter, ETag validation of the object is skipped.

$sel:bucket:S3Location', s3Location_bucket - The name of the Amazon S3 bucket where the application revision is stored.

$sel:key:S3Location', s3Location_key - The name of the Amazon S3 object that represents the bundled artifacts for the application revision.

$sel:version:S3Location', s3Location_version - A specific version of the Amazon S3 object that represents the bundled artifacts for the application revision.

If the version is not specified, the system uses the most recent version by default.

s3Location_bundleType :: Lens' S3Location (Maybe BundleType)

The file type of the application revision. Must be one of the following:

tar: A tar archive file.

tgz: A compressed tar archive file.

zip: A zip archive file.

s3Location_eTag :: Lens' S3Location (Maybe Text)

The ETag of the Amazon S3 object that represents the bundled artifacts for the application revision.

If the ETag is not specified as an input parameter, ETag validation of the object is skipped.

s3Location_bucket :: Lens' S3Location (Maybe Text)

The name of the Amazon S3 bucket where the application revision is stored.

s3Location_key :: Lens' S3Location (Maybe Text)

The name of the Amazon S3 object that represents the bundled artifacts for the application revision.

s3Location_version :: Lens' S3Location (Maybe Text)

A specific version of the Amazon S3 object that represents the bundled artifacts for the application revision.

If the version is not specified, the system uses the most recent version by default.