Network.AWS.CodeDeploy.Types.TargetGroupPairInfo

Description

Documentation

data TargetGroupPairInfo

Information about two target groups and how traffic is routed during an Amazon ECS deployment. An optional test traffic route can be specified.

See: newTargetGroupPairInfo smart constructor.

Constructors

TargetGroupPairInfo'

Fields

targetGroups :: Maybe [TargetGroupInfo]

One pair of target groups. One is associated with the original task set. The second is associated with the task set that serves traffic after the deployment is complete.

prodTrafficRoute :: Maybe TrafficRoute

The path used by a load balancer to route production traffic when an Amazon ECS deployment is complete.

testTrafficRoute :: Maybe TrafficRoute

An optional path used by a load balancer to route test traffic after an Amazon ECS deployment. Validation can occur while test traffic is served during a deployment.

newTargetGroupPairInfo :: TargetGroupPairInfo

Create a value of TargetGroupPairInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetGroups:TargetGroupPairInfo', targetGroupPairInfo_targetGroups - One pair of target groups. One is associated with the original task set. The second is associated with the task set that serves traffic after the deployment is complete.

$sel:prodTrafficRoute:TargetGroupPairInfo', targetGroupPairInfo_prodTrafficRoute - The path used by a load balancer to route production traffic when an Amazon ECS deployment is complete.

$sel:testTrafficRoute:TargetGroupPairInfo', targetGroupPairInfo_testTrafficRoute - An optional path used by a load balancer to route test traffic after an Amazon ECS deployment. Validation can occur while test traffic is served during a deployment.

targetGroupPairInfo_targetGroups :: Lens' TargetGroupPairInfo (Maybe [TargetGroupInfo])

One pair of target groups. One is associated with the original task set. The second is associated with the task set that serves traffic after the deployment is complete.

targetGroupPairInfo_prodTrafficRoute :: Lens' TargetGroupPairInfo (Maybe TrafficRoute)

The path used by a load balancer to route production traffic when an Amazon ECS deployment is complete.

targetGroupPairInfo_testTrafficRoute :: Lens' TargetGroupPairInfo (Maybe TrafficRoute)

An optional path used by a load balancer to route test traffic after an Amazon ECS deployment. Validation can occur while test traffic is served during a deployment.