Network.AWS.CodeDeploy.UpdateDeploymentGroup

Description

Changes information about a deployment group.

Creating a Request

data UpdateDeploymentGroup

Represents the input of an UpdateDeploymentGroup operation.

See: newUpdateDeploymentGroup smart constructor.

Constructors

UpdateDeploymentGroup'

Fields

serviceRoleArn :: Maybe Text

A replacement ARN for the service role, if you want to change it.

ec2TagSet :: Maybe EC2TagSet

Information about groups of tags applied to on-premises instances. The deployment group includes only EC2 instances identified by all the tag groups.

deploymentConfigName :: Maybe Text

The replacement deployment configuration name to use, if you want to change it.

onPremisesTagSet :: Maybe OnPremisesTagSet

Information about an on-premises instance tag set. The deployment group includes only on-premises instances identified by all the tag groups.

newDeploymentGroupName' :: Maybe Text

The new name of the deployment group, if you want to change it.

ec2TagFilters :: Maybe [EC2TagFilter]

The replacement set of Amazon EC2 tags on which to filter, if you want to change them. To keep the existing tags, enter their names. To remove tags, do not enter any tag names.

ecsServices :: Maybe [ECSService]

The target Amazon ECS services in the deployment group. This applies only to deployment groups that use the Amazon ECS compute platform. A target Amazon ECS service is specified as an Amazon ECS cluster and service name pair using the format <clustername>:<servicename>.

blueGreenDeploymentConfiguration :: Maybe BlueGreenDeploymentConfiguration

Information about blue/green deployment options for a deployment group.

loadBalancerInfo :: Maybe LoadBalancerInfo

Information about the load balancer used in a deployment.

outdatedInstancesStrategy :: Maybe OutdatedInstancesStrategy

Indicates what happens when new EC2 instances are launched mid-deployment and do not receive the deployed application revision.

If this option is set to UPDATE or is unspecified, CodeDeploy initiates one or more 'auto-update outdated instances' deployments to apply the deployed application revision to the new EC2 instances.

If this option is set to IGNORE, CodeDeploy does not initiate a deployment to update the new EC2 instances. This may result in instances having different revisions.

onPremisesInstanceTagFilters :: Maybe [TagFilter]

The replacement set of on-premises instance tags on which to filter, if you want to change them. To keep the existing tags, enter their names. To remove tags, do not enter any tag names.

alarmConfiguration :: Maybe AlarmConfiguration

Information to add or change about Amazon CloudWatch alarms when the deployment group is updated.

triggerConfigurations :: Maybe [TriggerConfig]

Information about triggers to change when the deployment group is updated. For examples, see Edit a Trigger in a CodeDeploy Deployment Group in the AWS CodeDeploy User Guide.

autoScalingGroups :: Maybe [Text]

The replacement list of Auto Scaling groups to be included in the deployment group, if you want to change them. To keep the Auto Scaling groups, enter their names. To remove Auto Scaling groups, do not enter any Auto Scaling group names.

deploymentStyle :: Maybe DeploymentStyle

Information about the type of deployment, either in-place or blue/green, you want to run and whether to route deployment traffic behind a load balancer.

autoRollbackConfiguration :: Maybe AutoRollbackConfiguration

Information for an automatic rollback configuration that is added or changed when a deployment group is updated.

applicationName :: Text

The application name that corresponds to the deployment group to update.

currentDeploymentGroupName :: Text

The current name of the deployment group.

newUpdateDeploymentGroup

Arguments

:: Text
$sel:applicationName:UpdateDeploymentGroup'

-> Text
$sel:currentDeploymentGroupName:UpdateDeploymentGroup'

-> UpdateDeploymentGroup

Create a value of UpdateDeploymentGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceRoleArn:UpdateDeploymentGroup', updateDeploymentGroup_serviceRoleArn - A replacement ARN for the service role, if you want to change it.

$sel:ec2TagSet:UpdateDeploymentGroup', updateDeploymentGroup_ec2TagSet - Information about groups of tags applied to on-premises instances. The deployment group includes only EC2 instances identified by all the tag groups.

$sel:deploymentConfigName:UpdateDeploymentGroup', updateDeploymentGroup_deploymentConfigName - The replacement deployment configuration name to use, if you want to change it.

$sel:onPremisesTagSet:UpdateDeploymentGroup', updateDeploymentGroup_onPremisesTagSet - Information about an on-premises instance tag set. The deployment group includes only on-premises instances identified by all the tag groups.

$sel:newDeploymentGroupName':UpdateDeploymentGroup', updateDeploymentGroup_newDeploymentGroupName - The new name of the deployment group, if you want to change it.

$sel:ec2TagFilters:UpdateDeploymentGroup', updateDeploymentGroup_ec2TagFilters - The replacement set of Amazon EC2 tags on which to filter, if you want to change them. To keep the existing tags, enter their names. To remove tags, do not enter any tag names.

$sel:ecsServices:UpdateDeploymentGroup', updateDeploymentGroup_ecsServices - The target Amazon ECS services in the deployment group. This applies only to deployment groups that use the Amazon ECS compute platform. A target Amazon ECS service is specified as an Amazon ECS cluster and service name pair using the format <clustername>:<servicename>.

$sel:blueGreenDeploymentConfiguration:UpdateDeploymentGroup', updateDeploymentGroup_blueGreenDeploymentConfiguration - Information about blue/green deployment options for a deployment group.

$sel:loadBalancerInfo:UpdateDeploymentGroup', updateDeploymentGroup_loadBalancerInfo - Information about the load balancer used in a deployment.

$sel:outdatedInstancesStrategy:UpdateDeploymentGroup', updateDeploymentGroup_outdatedInstancesStrategy - Indicates what happens when new EC2 instances are launched mid-deployment and do not receive the deployed application revision.

If this option is set to UPDATE or is unspecified, CodeDeploy initiates one or more 'auto-update outdated instances' deployments to apply the deployed application revision to the new EC2 instances.

If this option is set to IGNORE, CodeDeploy does not initiate a deployment to update the new EC2 instances. This may result in instances having different revisions.

$sel:onPremisesInstanceTagFilters:UpdateDeploymentGroup', updateDeploymentGroup_onPremisesInstanceTagFilters - The replacement set of on-premises instance tags on which to filter, if you want to change them. To keep the existing tags, enter their names. To remove tags, do not enter any tag names.

$sel:alarmConfiguration:UpdateDeploymentGroup', updateDeploymentGroup_alarmConfiguration - Information to add or change about Amazon CloudWatch alarms when the deployment group is updated.

$sel:triggerConfigurations:UpdateDeploymentGroup', updateDeploymentGroup_triggerConfigurations - Information about triggers to change when the deployment group is updated. For examples, see Edit a Trigger in a CodeDeploy Deployment Group in the AWS CodeDeploy User Guide.

$sel:autoScalingGroups:UpdateDeploymentGroup', updateDeploymentGroup_autoScalingGroups - The replacement list of Auto Scaling groups to be included in the deployment group, if you want to change them. To keep the Auto Scaling groups, enter their names. To remove Auto Scaling groups, do not enter any Auto Scaling group names.

$sel:deploymentStyle:UpdateDeploymentGroup', updateDeploymentGroup_deploymentStyle - Information about the type of deployment, either in-place or blue/green, you want to run and whether to route deployment traffic behind a load balancer.

$sel:autoRollbackConfiguration:UpdateDeploymentGroup', updateDeploymentGroup_autoRollbackConfiguration - Information for an automatic rollback configuration that is added or changed when a deployment group is updated.

$sel:applicationName:UpdateDeploymentGroup', updateDeploymentGroup_applicationName - The application name that corresponds to the deployment group to update.

$sel:currentDeploymentGroupName:UpdateDeploymentGroup', updateDeploymentGroup_currentDeploymentGroupName - The current name of the deployment group.

Request Lenses

updateDeploymentGroup_serviceRoleArn :: Lens' UpdateDeploymentGroup (Maybe Text)

A replacement ARN for the service role, if you want to change it.

updateDeploymentGroup_ec2TagSet :: Lens' UpdateDeploymentGroup (Maybe EC2TagSet)

Information about groups of tags applied to on-premises instances. The deployment group includes only EC2 instances identified by all the tag groups.

updateDeploymentGroup_deploymentConfigName :: Lens' UpdateDeploymentGroup (Maybe Text)

The replacement deployment configuration name to use, if you want to change it.

updateDeploymentGroup_onPremisesTagSet :: Lens' UpdateDeploymentGroup (Maybe OnPremisesTagSet)

Information about an on-premises instance tag set. The deployment group includes only on-premises instances identified by all the tag groups.

updateDeploymentGroup_newDeploymentGroupName :: Lens' UpdateDeploymentGroup (Maybe Text)

The new name of the deployment group, if you want to change it.

updateDeploymentGroup_ec2TagFilters :: Lens' UpdateDeploymentGroup (Maybe [EC2TagFilter])

The replacement set of Amazon EC2 tags on which to filter, if you want to change them. To keep the existing tags, enter their names. To remove tags, do not enter any tag names.

updateDeploymentGroup_ecsServices :: Lens' UpdateDeploymentGroup (Maybe [ECSService])

The target Amazon ECS services in the deployment group. This applies only to deployment groups that use the Amazon ECS compute platform. A target Amazon ECS service is specified as an Amazon ECS cluster and service name pair using the format <clustername>:<servicename>.

updateDeploymentGroup_blueGreenDeploymentConfiguration :: Lens' UpdateDeploymentGroup (Maybe BlueGreenDeploymentConfiguration)

Information about blue/green deployment options for a deployment group.

updateDeploymentGroup_loadBalancerInfo :: Lens' UpdateDeploymentGroup (Maybe LoadBalancerInfo)

Information about the load balancer used in a deployment.

updateDeploymentGroup_outdatedInstancesStrategy :: Lens' UpdateDeploymentGroup (Maybe OutdatedInstancesStrategy)

Indicates what happens when new EC2 instances are launched mid-deployment and do not receive the deployed application revision.

If this option is set to UPDATE or is unspecified, CodeDeploy initiates one or more 'auto-update outdated instances' deployments to apply the deployed application revision to the new EC2 instances.

If this option is set to IGNORE, CodeDeploy does not initiate a deployment to update the new EC2 instances. This may result in instances having different revisions.

updateDeploymentGroup_onPremisesInstanceTagFilters :: Lens' UpdateDeploymentGroup (Maybe [TagFilter])

The replacement set of on-premises instance tags on which to filter, if you want to change them. To keep the existing tags, enter their names. To remove tags, do not enter any tag names.

updateDeploymentGroup_alarmConfiguration :: Lens' UpdateDeploymentGroup (Maybe AlarmConfiguration)

Information to add or change about Amazon CloudWatch alarms when the deployment group is updated.

updateDeploymentGroup_triggerConfigurations :: Lens' UpdateDeploymentGroup (Maybe [TriggerConfig])

Information about triggers to change when the deployment group is updated. For examples, see Edit a Trigger in a CodeDeploy Deployment Group in the AWS CodeDeploy User Guide.

updateDeploymentGroup_autoScalingGroups :: Lens' UpdateDeploymentGroup (Maybe [Text])

The replacement list of Auto Scaling groups to be included in the deployment group, if you want to change them. To keep the Auto Scaling groups, enter their names. To remove Auto Scaling groups, do not enter any Auto Scaling group names.

updateDeploymentGroup_deploymentStyle :: Lens' UpdateDeploymentGroup (Maybe DeploymentStyle)

Information about the type of deployment, either in-place or blue/green, you want to run and whether to route deployment traffic behind a load balancer.

updateDeploymentGroup_autoRollbackConfiguration :: Lens' UpdateDeploymentGroup (Maybe AutoRollbackConfiguration)

Information for an automatic rollback configuration that is added or changed when a deployment group is updated.

updateDeploymentGroup_applicationName :: Lens' UpdateDeploymentGroup Text

The application name that corresponds to the deployment group to update.

updateDeploymentGroup_currentDeploymentGroupName :: Lens' UpdateDeploymentGroup Text

The current name of the deployment group.

Destructuring the Response

data UpdateDeploymentGroupResponse

Represents the output of an UpdateDeploymentGroup operation.

See: newUpdateDeploymentGroupResponse smart constructor.

Constructors

UpdateDeploymentGroupResponse'

Fields

hooksNotCleanedUp :: Maybe [AutoScalingGroup]

If the output contains no data, and the corresponding deployment group contained at least one Auto Scaling group, AWS CodeDeploy successfully removed all corresponding Auto Scaling lifecycle event hooks from the AWS account. If the output contains data, AWS CodeDeploy could not remove some Auto Scaling lifecycle event hooks from the AWS account.

httpStatus :: Int

The response's http status code.

newUpdateDeploymentGroupResponse

Arguments

:: Int
$sel:httpStatus:UpdateDeploymentGroupResponse'

-> UpdateDeploymentGroupResponse

Create a value of UpdateDeploymentGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hooksNotCleanedUp:UpdateDeploymentGroupResponse', updateDeploymentGroupResponse_hooksNotCleanedUp - If the output contains no data, and the corresponding deployment group contained at least one Auto Scaling group, AWS CodeDeploy successfully removed all corresponding Auto Scaling lifecycle event hooks from the AWS account. If the output contains data, AWS CodeDeploy could not remove some Auto Scaling lifecycle event hooks from the AWS account.

$sel:httpStatus:UpdateDeploymentGroupResponse', updateDeploymentGroupResponse_httpStatus - The response's http status code.

Response Lenses

updateDeploymentGroupResponse_hooksNotCleanedUp :: Lens' UpdateDeploymentGroupResponse (Maybe [AutoScalingGroup])

If the output contains no data, and the corresponding deployment group contained at least one Auto Scaling group, AWS CodeDeploy successfully removed all corresponding Auto Scaling lifecycle event hooks from the AWS account. If the output contains data, AWS CodeDeploy could not remove some Auto Scaling lifecycle event hooks from the AWS account.

updateDeploymentGroupResponse_httpStatus :: Lens' UpdateDeploymentGroupResponse Int

The response's http status code.