Network.AWS.CodeGuruProfiler.BatchGetFrameMetricData

Description

Returns the time series of values for a requested list of frame metrics from a time period.

Creating a Request

data BatchGetFrameMetricData

The structure representing the BatchGetFrameMetricDataRequest.

See: newBatchGetFrameMetricData smart constructor.

Constructors

BatchGetFrameMetricData'

Fields

targetResolution :: Maybe AggregationPeriod

The requested resolution of time steps for the returned time series of values. If the requested target resolution is not available due to data not being retained we provide a best effort result by falling back to the most granular available resolution after the target resolution. There are 3 valid values.

P1D — 1 day

PT1H — 1 hour

PT5M — 5 minutes

frameMetrics :: Maybe [FrameMetric]

The details of the metrics that are used to request a time series of values. The metric includes the name of the frame, the aggregation type to calculate the metric value for the frame, and the thread states to use to get the count for the metric value of the frame.

startTime :: Maybe POSIX

The start time of the time period for the frame metrics used to return the time series values. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

period :: Maybe Text

The duration of the frame metrics used to return the time series values. Specify using the ISO 8601 format. The maximum period duration is one day (PT24H or P1D).

endTime :: Maybe POSIX

The end time of the time period for the returned time series values. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

profilingGroupName :: Text

The name of the profiling group associated with the the frame metrics used to return the time series values.

newBatchGetFrameMetricData

Arguments

:: Text
$sel:profilingGroupName:BatchGetFrameMetricData'

-> BatchGetFrameMetricData

Create a value of BatchGetFrameMetricData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetResolution:BatchGetFrameMetricData', batchGetFrameMetricData_targetResolution - The requested resolution of time steps for the returned time series of values. If the requested target resolution is not available due to data not being retained we provide a best effort result by falling back to the most granular available resolution after the target resolution. There are 3 valid values.

P1D — 1 day

PT1H — 1 hour

PT5M — 5 minutes

$sel:frameMetrics:BatchGetFrameMetricData', batchGetFrameMetricData_frameMetrics - The details of the metrics that are used to request a time series of values. The metric includes the name of the frame, the aggregation type to calculate the metric value for the frame, and the thread states to use to get the count for the metric value of the frame.

$sel:startTime:BatchGetFrameMetricData', batchGetFrameMetricData_startTime - The start time of the time period for the frame metrics used to return the time series values. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

$sel:period:BatchGetFrameMetricData', batchGetFrameMetricData_period - The duration of the frame metrics used to return the time series values. Specify using the ISO 8601 format. The maximum period duration is one day (PT24H or P1D).

$sel:endTime:BatchGetFrameMetricData', batchGetFrameMetricData_endTime - The end time of the time period for the returned time series values. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

$sel:profilingGroupName:BatchGetFrameMetricData', batchGetFrameMetricData_profilingGroupName - The name of the profiling group associated with the the frame metrics used to return the time series values.

Request Lenses

batchGetFrameMetricData_targetResolution :: Lens' BatchGetFrameMetricData (Maybe AggregationPeriod)

The requested resolution of time steps for the returned time series of values. If the requested target resolution is not available due to data not being retained we provide a best effort result by falling back to the most granular available resolution after the target resolution. There are 3 valid values.

P1D — 1 day

PT1H — 1 hour

PT5M — 5 minutes

batchGetFrameMetricData_frameMetrics :: Lens' BatchGetFrameMetricData (Maybe [FrameMetric])

The details of the metrics that are used to request a time series of values. The metric includes the name of the frame, the aggregation type to calculate the metric value for the frame, and the thread states to use to get the count for the metric value of the frame.

batchGetFrameMetricData_startTime :: Lens' BatchGetFrameMetricData (Maybe UTCTime)

The start time of the time period for the frame metrics used to return the time series values. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

batchGetFrameMetricData_period :: Lens' BatchGetFrameMetricData (Maybe Text)

The duration of the frame metrics used to return the time series values. Specify using the ISO 8601 format. The maximum period duration is one day (PT24H or P1D).

batchGetFrameMetricData_endTime :: Lens' BatchGetFrameMetricData (Maybe UTCTime)

The end time of the time period for the returned time series values. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

batchGetFrameMetricData_profilingGroupName :: Lens' BatchGetFrameMetricData Text

The name of the profiling group associated with the the frame metrics used to return the time series values.

Destructuring the Response

data BatchGetFrameMetricDataResponse

The structure representing the BatchGetFrameMetricDataResponse.

See: newBatchGetFrameMetricDataResponse smart constructor.

Constructors

BatchGetFrameMetricDataResponse'

Fields

httpStatus :: Int

The response's http status code.

endTime :: POSIX

The end time of the time period for the returned time series values. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

endTimes :: [TimestampStructure]

List of instances, or time steps, in the time series. For example, if the period is one day (PT24H)), and the resolution is five minutes (PT5M), then there are 288 endTimes in the list that are each five minutes appart.

frameMetricData :: [FrameMetricDatum]

Details of the metrics to request a time series of values. The metric includes the name of the frame, the aggregation type to calculate the metric value for the frame, and the thread states to use to get the count for the metric value of the frame.

resolution :: AggregationPeriod

Resolution or granularity of the profile data used to generate the time series. This is the value used to jump through time steps in a time series. There are 3 valid values.

P1D — 1 day

PT1H — 1 hour

PT5M — 5 minutes

startTime :: POSIX

The start time of the time period for the returned time series values. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

unprocessedEndTimes :: HashMap Text [TimestampStructure]

List of instances which remained unprocessed. This will create a missing time step in the list of end times.

newBatchGetFrameMetricDataResponse

Arguments

:: Int
$sel:httpStatus:BatchGetFrameMetricDataResponse'

-> UTCTime
$sel:endTime:BatchGetFrameMetricData'

-> AggregationPeriod
$sel:resolution:BatchGetFrameMetricDataResponse'

-> UTCTime
$sel:startTime:BatchGetFrameMetricData'

-> BatchGetFrameMetricDataResponse

Create a value of BatchGetFrameMetricDataResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:BatchGetFrameMetricDataResponse', batchGetFrameMetricDataResponse_httpStatus - The response's http status code.

$sel:endTime:BatchGetFrameMetricData', batchGetFrameMetricDataResponse_endTime - The end time of the time period for the returned time series values. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

$sel:endTimes:BatchGetFrameMetricDataResponse', batchGetFrameMetricDataResponse_endTimes - List of instances, or time steps, in the time series. For example, if the period is one day (PT24H)), and the resolution is five minutes (PT5M), then there are 288 endTimes in the list that are each five minutes appart.

$sel:frameMetricData:BatchGetFrameMetricDataResponse', batchGetFrameMetricDataResponse_frameMetricData - Details of the metrics to request a time series of values. The metric includes the name of the frame, the aggregation type to calculate the metric value for the frame, and the thread states to use to get the count for the metric value of the frame.

$sel:resolution:BatchGetFrameMetricDataResponse', batchGetFrameMetricDataResponse_resolution - Resolution or granularity of the profile data used to generate the time series. This is the value used to jump through time steps in a time series. There are 3 valid values.

P1D — 1 day

PT1H — 1 hour

PT5M — 5 minutes

$sel:startTime:BatchGetFrameMetricData', batchGetFrameMetricDataResponse_startTime - The start time of the time period for the returned time series values. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

$sel:unprocessedEndTimes:BatchGetFrameMetricDataResponse', batchGetFrameMetricDataResponse_unprocessedEndTimes - List of instances which remained unprocessed. This will create a missing time step in the list of end times.

Response Lenses

batchGetFrameMetricDataResponse_httpStatus :: Lens' BatchGetFrameMetricDataResponse Int

The response's http status code.

batchGetFrameMetricDataResponse_endTime :: Lens' BatchGetFrameMetricDataResponse UTCTime

The end time of the time period for the returned time series values. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

batchGetFrameMetricDataResponse_endTimes :: Lens' BatchGetFrameMetricDataResponse [TimestampStructure]

List of instances, or time steps, in the time series. For example, if the period is one day (PT24H)), and the resolution is five minutes (PT5M), then there are 288 endTimes in the list that are each five minutes appart.

batchGetFrameMetricDataResponse_frameMetricData :: Lens' BatchGetFrameMetricDataResponse [FrameMetricDatum]

Details of the metrics to request a time series of values. The metric includes the name of the frame, the aggregation type to calculate the metric value for the frame, and the thread states to use to get the count for the metric value of the frame.

batchGetFrameMetricDataResponse_resolution :: Lens' BatchGetFrameMetricDataResponse AggregationPeriod

Resolution or granularity of the profile data used to generate the time series. This is the value used to jump through time steps in a time series. There are 3 valid values.

P1D — 1 day

PT1H — 1 hour

PT5M — 5 minutes

batchGetFrameMetricDataResponse_startTime :: Lens' BatchGetFrameMetricDataResponse UTCTime

The start time of the time period for the returned time series values. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

batchGetFrameMetricDataResponse_unprocessedEndTimes :: Lens' BatchGetFrameMetricDataResponse (HashMap Text [TimestampStructure])

List of instances which remained unprocessed. This will create a missing time step in the list of end times.