Network.AWS.CodeGuruProfiler.ConfigureAgent

Description

Used by profiler agents to report their current state and to receive remote configuration updates. For example, ConfigureAgent can be used to tell an agent whether to profile or not and for how long to return profiling data.

Creating a Request

data ConfigureAgent

The structure representing the configureAgentRequest.

See: newConfigureAgent smart constructor.

Constructors

ConfigureAgent'

Fields

fleetInstanceId :: Maybe Text

A universally unique identifier (UUID) for a profiling instance. For example, if the profiling instance is an Amazon EC2 instance, it is the instance ID. If it is an AWS Fargate container, it is the container's task ID.

metadata :: Maybe (HashMap MetadataField Text)

Metadata captured about the compute platform the agent is running on. It includes information about sampling and reporting. The valid fields are:

COMPUTE_PLATFORM - The compute platform on which the agent is running

AGENT_ID - The ID for an agent instance.

AWS_REQUEST_ID - The AWS request ID of a Lambda invocation.

EXECUTION_ENVIRONMENT - The execution environment a Lambda function is running on.

LAMBDA_FUNCTION_ARN - The Amazon Resource Name (ARN) that is used to invoke a Lambda function.

LAMBDA_MEMORY_LIMIT_IN_MB - The memory allocated to a Lambda function.

LAMBDA_REMAINING_TIME_IN_MILLISECONDS - The time in milliseconds before execution of a Lambda function times out.

LAMBDA_TIME_GAP_BETWEEN_INVOKES_IN_MILLISECONDS - The time in milliseconds between two invocations of a Lambda function.

LAMBDA_PREVIOUS_EXECUTION_TIME_IN_MILLISECONDS - The time in milliseconds for the previous Lambda invocation.

profilingGroupName :: Text

The name of the profiling group for which the configured agent is collecting profiling data.

newConfigureAgent

Arguments

:: Text
$sel:profilingGroupName:ConfigureAgent'

-> ConfigureAgent

Create a value of ConfigureAgent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fleetInstanceId:ConfigureAgent', configureAgent_fleetInstanceId - A universally unique identifier (UUID) for a profiling instance. For example, if the profiling instance is an Amazon EC2 instance, it is the instance ID. If it is an AWS Fargate container, it is the container's task ID.

$sel:metadata:ConfigureAgent', configureAgent_metadata - Metadata captured about the compute platform the agent is running on. It includes information about sampling and reporting. The valid fields are:

COMPUTE_PLATFORM - The compute platform on which the agent is running

AGENT_ID - The ID for an agent instance.

AWS_REQUEST_ID - The AWS request ID of a Lambda invocation.

EXECUTION_ENVIRONMENT - The execution environment a Lambda function is running on.

LAMBDA_FUNCTION_ARN - The Amazon Resource Name (ARN) that is used to invoke a Lambda function.

LAMBDA_MEMORY_LIMIT_IN_MB - The memory allocated to a Lambda function.

LAMBDA_REMAINING_TIME_IN_MILLISECONDS - The time in milliseconds before execution of a Lambda function times out.

LAMBDA_TIME_GAP_BETWEEN_INVOKES_IN_MILLISECONDS - The time in milliseconds between two invocations of a Lambda function.

LAMBDA_PREVIOUS_EXECUTION_TIME_IN_MILLISECONDS - The time in milliseconds for the previous Lambda invocation.

$sel:profilingGroupName:ConfigureAgent', configureAgent_profilingGroupName - The name of the profiling group for which the configured agent is collecting profiling data.

Request Lenses

configureAgent_fleetInstanceId :: Lens' ConfigureAgent (Maybe Text)

A universally unique identifier (UUID) for a profiling instance. For example, if the profiling instance is an Amazon EC2 instance, it is the instance ID. If it is an AWS Fargate container, it is the container's task ID.

configureAgent_metadata :: Lens' ConfigureAgent (Maybe (HashMap MetadataField Text))

Metadata captured about the compute platform the agent is running on. It includes information about sampling and reporting. The valid fields are:

COMPUTE_PLATFORM - The compute platform on which the agent is running

AGENT_ID - The ID for an agent instance.

AWS_REQUEST_ID - The AWS request ID of a Lambda invocation.

EXECUTION_ENVIRONMENT - The execution environment a Lambda function is running on.

LAMBDA_FUNCTION_ARN - The Amazon Resource Name (ARN) that is used to invoke a Lambda function.

LAMBDA_MEMORY_LIMIT_IN_MB - The memory allocated to a Lambda function.

LAMBDA_REMAINING_TIME_IN_MILLISECONDS - The time in milliseconds before execution of a Lambda function times out.

LAMBDA_TIME_GAP_BETWEEN_INVOKES_IN_MILLISECONDS - The time in milliseconds between two invocations of a Lambda function.

LAMBDA_PREVIOUS_EXECUTION_TIME_IN_MILLISECONDS - The time in milliseconds for the previous Lambda invocation.

configureAgent_profilingGroupName :: Lens' ConfigureAgent Text

The name of the profiling group for which the configured agent is collecting profiling data.

Destructuring the Response

data ConfigureAgentResponse

The structure representing the configureAgentResponse.

See: newConfigureAgentResponse smart constructor.

Constructors

ConfigureAgentResponse'

Fields

httpStatus :: Int

The response's http status code.

configuration :: AgentConfiguration

An AgentConfiguration object that specifies if an agent profiles or not and for how long to return profiling data.

newConfigureAgentResponse

Arguments

:: Int
$sel:httpStatus:ConfigureAgentResponse'

-> AgentConfiguration
$sel:configuration:ConfigureAgentResponse'

-> ConfigureAgentResponse

Create a value of ConfigureAgentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:ConfigureAgentResponse', configureAgentResponse_httpStatus - The response's http status code.

$sel:configuration:ConfigureAgentResponse', configureAgentResponse_configuration - An AgentConfiguration object that specifies if an agent profiles or not and for how long to return profiling data.

Response Lenses

configureAgentResponse_httpStatus :: Lens' ConfigureAgentResponse Int

The response's http status code.

configureAgentResponse_configuration :: Lens' ConfigureAgentResponse AgentConfiguration

An AgentConfiguration object that specifies if an agent profiles or not and for how long to return profiling data.