Network.AWS.CodeGuruProfiler.CreateProfilingGroup

Description

Creates a profiling group.

Creating a Request

data CreateProfilingGroup

The structure representing the createProfiliingGroupRequest.

See: newCreateProfilingGroup smart constructor.

Constructors

CreateProfilingGroup'

Fields

computePlatform :: Maybe ComputePlatform

The compute platform of the profiling group. Use AWSLambda if your application runs on AWS Lambda. Use Default if your application runs on a compute platform that is not AWS Lambda, such an Amazon EC2 instance, an on-premises server, or a different platform. If not specified, Default is used.

agentOrchestrationConfig :: Maybe AgentOrchestrationConfig

Specifies whether profiling is enabled or disabled for the created profiling group.

tags :: Maybe (HashMap Text Text)

A list of tags to add to the created profiling group.

clientToken :: Text

Amazon CodeGuru Profiler uses this universally unique identifier (UUID) to prevent the accidental creation of duplicate profiling groups if there are failures and retries.

profilingGroupName :: Text

The name of the profiling group to create.

newCreateProfilingGroup

Arguments

:: Text
$sel:clientToken:CreateProfilingGroup'

-> Text
$sel:profilingGroupName:CreateProfilingGroup'

-> CreateProfilingGroup

Create a value of CreateProfilingGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:computePlatform:CreateProfilingGroup', createProfilingGroup_computePlatform - The compute platform of the profiling group. Use AWSLambda if your application runs on AWS Lambda. Use Default if your application runs on a compute platform that is not AWS Lambda, such an Amazon EC2 instance, an on-premises server, or a different platform. If not specified, Default is used.

$sel:agentOrchestrationConfig:CreateProfilingGroup', createProfilingGroup_agentOrchestrationConfig - Specifies whether profiling is enabled or disabled for the created profiling group.

$sel:tags:CreateProfilingGroup', createProfilingGroup_tags - A list of tags to add to the created profiling group.

$sel:clientToken:CreateProfilingGroup', createProfilingGroup_clientToken - Amazon CodeGuru Profiler uses this universally unique identifier (UUID) to prevent the accidental creation of duplicate profiling groups if there are failures and retries.

$sel:profilingGroupName:CreateProfilingGroup', createProfilingGroup_profilingGroupName - The name of the profiling group to create.

Request Lenses

createProfilingGroup_computePlatform :: Lens' CreateProfilingGroup (Maybe ComputePlatform)

The compute platform of the profiling group. Use AWSLambda if your application runs on AWS Lambda. Use Default if your application runs on a compute platform that is not AWS Lambda, such an Amazon EC2 instance, an on-premises server, or a different platform. If not specified, Default is used.

createProfilingGroup_agentOrchestrationConfig :: Lens' CreateProfilingGroup (Maybe AgentOrchestrationConfig)

Specifies whether profiling is enabled or disabled for the created profiling group.

createProfilingGroup_tags :: Lens' CreateProfilingGroup (Maybe (HashMap Text Text))

A list of tags to add to the created profiling group.

createProfilingGroup_clientToken :: Lens' CreateProfilingGroup Text

Amazon CodeGuru Profiler uses this universally unique identifier (UUID) to prevent the accidental creation of duplicate profiling groups if there are failures and retries.

createProfilingGroup_profilingGroupName :: Lens' CreateProfilingGroup Text

The name of the profiling group to create.

Destructuring the Response

data CreateProfilingGroupResponse

The structure representing the createProfilingGroupResponse.

See: newCreateProfilingGroupResponse smart constructor.

Constructors

CreateProfilingGroupResponse'

Fields

httpStatus :: Int

The response's http status code.

profilingGroup :: ProfilingGroupDescription

The returned ProfilingGroupDescription object that contains information about the created profiling group.

newCreateProfilingGroupResponse

Arguments

:: Int
$sel:httpStatus:CreateProfilingGroupResponse'

-> ProfilingGroupDescription
$sel:profilingGroup:CreateProfilingGroupResponse'

-> CreateProfilingGroupResponse

Create a value of CreateProfilingGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateProfilingGroupResponse', createProfilingGroupResponse_httpStatus - The response's http status code.

$sel:profilingGroup:CreateProfilingGroupResponse', createProfilingGroupResponse_profilingGroup - The returned ProfilingGroupDescription object that contains information about the created profiling group.

Response Lenses

createProfilingGroupResponse_httpStatus :: Lens' CreateProfilingGroupResponse Int

The response's http status code.

createProfilingGroupResponse_profilingGroup :: Lens' CreateProfilingGroupResponse ProfilingGroupDescription

The returned ProfilingGroupDescription object that contains information about the created profiling group.