Network.AWS.CodeGuruProfiler.DescribeProfilingGroup

Description

Returns a ProfilingGroupDescription object that contains information about the requested profiling group.

Creating a Request

data DescribeProfilingGroup

The structure representing the describeProfilingGroupRequest.

See: newDescribeProfilingGroup smart constructor.

Constructors

DescribeProfilingGroup'

Fields

profilingGroupName :: Text

The name of the profiling group to get information about.

newDescribeProfilingGroup

Arguments

:: Text
$sel:profilingGroupName:DescribeProfilingGroup'

-> DescribeProfilingGroup

Create a value of DescribeProfilingGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:profilingGroupName:DescribeProfilingGroup', describeProfilingGroup_profilingGroupName - The name of the profiling group to get information about.

Request Lenses

describeProfilingGroup_profilingGroupName :: Lens' DescribeProfilingGroup Text

The name of the profiling group to get information about.

Destructuring the Response

data DescribeProfilingGroupResponse

The structure representing the describeProfilingGroupResponse.

See: newDescribeProfilingGroupResponse smart constructor.

Constructors

DescribeProfilingGroupResponse'

Fields

httpStatus :: Int

The response's http status code.

profilingGroup :: ProfilingGroupDescription

The returned ProfilingGroupDescription object that contains information about the requested profiling group.

newDescribeProfilingGroupResponse

Arguments

:: Int
$sel:httpStatus:DescribeProfilingGroupResponse'

-> ProfilingGroupDescription
$sel:profilingGroup:DescribeProfilingGroupResponse'

-> DescribeProfilingGroupResponse

Create a value of DescribeProfilingGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DescribeProfilingGroupResponse', describeProfilingGroupResponse_httpStatus - The response's http status code.

$sel:profilingGroup:DescribeProfilingGroupResponse', describeProfilingGroupResponse_profilingGroup - The returned ProfilingGroupDescription object that contains information about the requested profiling group.

Response Lenses

describeProfilingGroupResponse_httpStatus :: Lens' DescribeProfilingGroupResponse Int

The response's http status code.

describeProfilingGroupResponse_profilingGroup :: Lens' DescribeProfilingGroupResponse ProfilingGroupDescription

The returned ProfilingGroupDescription object that contains information about the requested profiling group.