Network.AWS.CodeGuruProfiler.GetRecommendations

Description

Returns a list of Recommendation objects that contain recommendations for a profiling group for a given time period. A list of Anomaly objects that contains details about anomalies detected in the profiling group for the same time period is also returned.

Creating a Request

data GetRecommendations

The structure representing the GetRecommendationsRequest.

See: newGetRecommendations smart constructor.

Constructors

GetRecommendations'

Fields

locale :: Maybe Text

The language used to provide analysis. Specify using a string that is one of the following BCP 47 language codes.

de-DE - German, Germany

en-GB - English, United Kingdom

en-US - English, United States

es-ES - Spanish, Spain

fr-FR - French, France

it-IT - Italian, Italy

ja-JP - Japanese, Japan

ko-KR - Korean, Republic of Korea

pt-BR - Portugese, Brazil

zh-CN - Chinese, China

zh-TW - Chinese, Taiwan

endTime :: POSIX

The start time of the profile to get analysis data about. You must specify startTime and endTime. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

profilingGroupName :: Text

The name of the profiling group to get analysis data about.

startTime :: POSIX

The end time of the profile to get analysis data about. You must specify startTime and endTime. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

newGetRecommendations

Arguments

:: UTCTime
$sel:endTime:GetRecommendations'

-> Text
$sel:profilingGroupName:GetRecommendations'

-> UTCTime
$sel:startTime:GetRecommendations'

-> GetRecommendations

Create a value of GetRecommendations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:locale:GetRecommendations', getRecommendations_locale - The language used to provide analysis. Specify using a string that is one of the following BCP 47 language codes.

de-DE - German, Germany

en-GB - English, United Kingdom

en-US - English, United States

es-ES - Spanish, Spain

fr-FR - French, France

it-IT - Italian, Italy

ja-JP - Japanese, Japan

ko-KR - Korean, Republic of Korea

pt-BR - Portugese, Brazil

zh-CN - Chinese, China

zh-TW - Chinese, Taiwan

$sel:endTime:GetRecommendations', getRecommendations_endTime - The start time of the profile to get analysis data about. You must specify startTime and endTime. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

$sel:profilingGroupName:GetRecommendations', getRecommendations_profilingGroupName - The name of the profiling group to get analysis data about.

$sel:startTime:GetRecommendations', getRecommendations_startTime - The end time of the profile to get analysis data about. You must specify startTime and endTime. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

Request Lenses

getRecommendations_locale :: Lens' GetRecommendations (Maybe Text)

The language used to provide analysis. Specify using a string that is one of the following BCP 47 language codes.

de-DE - German, Germany

en-GB - English, United Kingdom

en-US - English, United States

es-ES - Spanish, Spain

fr-FR - French, France

it-IT - Italian, Italy

ja-JP - Japanese, Japan

ko-KR - Korean, Republic of Korea

pt-BR - Portugese, Brazil

zh-CN - Chinese, China

zh-TW - Chinese, Taiwan

getRecommendations_endTime :: Lens' GetRecommendations UTCTime

The start time of the profile to get analysis data about. You must specify startTime and endTime. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

getRecommendations_profilingGroupName :: Lens' GetRecommendations Text

The name of the profiling group to get analysis data about.

getRecommendations_startTime :: Lens' GetRecommendations UTCTime

The end time of the profile to get analysis data about. You must specify startTime and endTime. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

Destructuring the Response

data GetRecommendationsResponse

The structure representing the GetRecommendationsResponse.

See: newGetRecommendationsResponse smart constructor.

Constructors

GetRecommendationsResponse'

Fields

httpStatus :: Int

The response's http status code.

anomalies :: [Anomaly]

The list of anomalies that the analysis has found for this profile.

profileEndTime :: POSIX

The end time of the profile the analysis data is about. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

profileStartTime :: POSIX

The start time of the profile the analysis data is about. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

profilingGroupName :: Text

The name of the profiling group the analysis data is about.

recommendations :: [Recommendation]

The list of recommendations that the analysis found for this profile.

newGetRecommendationsResponse

Arguments

:: Int
$sel:httpStatus:GetRecommendationsResponse'

-> UTCTime
$sel:profileEndTime:GetRecommendationsResponse'

-> UTCTime
$sel:profileStartTime:GetRecommendationsResponse'

-> Text
$sel:profilingGroupName:GetRecommendations'

-> GetRecommendationsResponse

Create a value of GetRecommendationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetRecommendationsResponse', getRecommendationsResponse_httpStatus - The response's http status code.

$sel:anomalies:GetRecommendationsResponse', getRecommendationsResponse_anomalies - The list of anomalies that the analysis has found for this profile.

$sel:profileEndTime:GetRecommendationsResponse', getRecommendationsResponse_profileEndTime - The end time of the profile the analysis data is about. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

$sel:profileStartTime:GetRecommendationsResponse', getRecommendationsResponse_profileStartTime - The start time of the profile the analysis data is about. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

$sel:profilingGroupName:GetRecommendations', getRecommendationsResponse_profilingGroupName - The name of the profiling group the analysis data is about.

$sel:recommendations:GetRecommendationsResponse', getRecommendationsResponse_recommendations - The list of recommendations that the analysis found for this profile.

Response Lenses

getRecommendationsResponse_httpStatus :: Lens' GetRecommendationsResponse Int

The response's http status code.

getRecommendationsResponse_anomalies :: Lens' GetRecommendationsResponse [Anomaly]

The list of anomalies that the analysis has found for this profile.

getRecommendationsResponse_profileEndTime :: Lens' GetRecommendationsResponse UTCTime

The end time of the profile the analysis data is about. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

getRecommendationsResponse_profileStartTime :: Lens' GetRecommendationsResponse UTCTime

The start time of the profile the analysis data is about. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

getRecommendationsResponse_profilingGroupName :: Lens' GetRecommendationsResponse Text

The name of the profiling group the analysis data is about.

getRecommendationsResponse_recommendations :: Lens' GetRecommendationsResponse [Recommendation]

The list of recommendations that the analysis found for this profile.