Network.AWS.CodeGuruProfiler.Lens

Description

Operations

GetRecommendations

getRecommendations_locale :: Lens' GetRecommendations (Maybe Text)

The language used to provide analysis. Specify using a string that is one of the following BCP 47 language codes.

de-DE - German, Germany

en-GB - English, United Kingdom

en-US - English, United States

es-ES - Spanish, Spain

fr-FR - French, France

it-IT - Italian, Italy

ja-JP - Japanese, Japan

ko-KR - Korean, Republic of Korea

pt-BR - Portugese, Brazil

zh-CN - Chinese, China

zh-TW - Chinese, Taiwan

getRecommendations_endTime :: Lens' GetRecommendations UTCTime

The start time of the profile to get analysis data about. You must specify startTime and endTime. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

getRecommendations_profilingGroupName :: Lens' GetRecommendations Text

The name of the profiling group to get analysis data about.

getRecommendations_startTime :: Lens' GetRecommendations UTCTime

The end time of the profile to get analysis data about. You must specify startTime and endTime. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

getRecommendationsResponse_httpStatus :: Lens' GetRecommendationsResponse Int

The response's http status code.

getRecommendationsResponse_anomalies :: Lens' GetRecommendationsResponse [Anomaly]

The list of anomalies that the analysis has found for this profile.

getRecommendationsResponse_profileEndTime :: Lens' GetRecommendationsResponse UTCTime

The end time of the profile the analysis data is about. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

getRecommendationsResponse_profileStartTime :: Lens' GetRecommendationsResponse UTCTime

The start time of the profile the analysis data is about. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

getRecommendationsResponse_profilingGroupName :: Lens' GetRecommendationsResponse Text

The name of the profiling group the analysis data is about.

getRecommendationsResponse_recommendations :: Lens' GetRecommendationsResponse [Recommendation]

The list of recommendations that the analysis found for this profile.

AddNotificationChannels

addNotificationChannels_channels :: Lens' AddNotificationChannels (NonEmpty Channel)

One or 2 channels to report to when anomalies are detected.

addNotificationChannels_profilingGroupName :: Lens' AddNotificationChannels Text

The name of the profiling group that we are setting up notifications for.

addNotificationChannelsResponse_notificationConfiguration :: Lens' AddNotificationChannelsResponse (Maybe NotificationConfiguration)

The new notification configuration for this profiling group.

addNotificationChannelsResponse_httpStatus :: Lens' AddNotificationChannelsResponse Int

The response's http status code.

DescribeProfilingGroup

describeProfilingGroup_profilingGroupName :: Lens' DescribeProfilingGroup Text

The name of the profiling group to get information about.

describeProfilingGroupResponse_httpStatus :: Lens' DescribeProfilingGroupResponse Int

The response's http status code.

describeProfilingGroupResponse_profilingGroup :: Lens' DescribeProfilingGroupResponse ProfilingGroupDescription

The returned ProfilingGroupDescription object that contains information about the requested profiling group.

PutPermission

putPermission_revisionId :: Lens' PutPermission (Maybe Text)

A universally unique identifier (UUID) for the revision of the policy you are adding to the profiling group. Do not specify this when you add permissions to a profiling group for the first time. If a policy already exists on the profiling group, you must specify the revisionId.

putPermission_actionGroup :: Lens' PutPermission ActionGroup

Specifies an action group that contains permissions to add to a profiling group resource. One action group is supported, agentPermissions, which grants permission to perform actions required by the profiling agent, ConfigureAgent and PostAgentProfile permissions.

putPermission_principals :: Lens' PutPermission (NonEmpty Text)

A list ARNs for the roles and users you want to grant access to the profiling group. Wildcards are not are supported in the ARNs.

putPermission_profilingGroupName :: Lens' PutPermission Text

The name of the profiling group to grant access to.

putPermissionResponse_httpStatus :: Lens' PutPermissionResponse Int

The response's http status code.

putPermissionResponse_policy :: Lens' PutPermissionResponse Text

The JSON-formatted resource-based policy on the profiling group that includes the added permissions.

putPermissionResponse_revisionId :: Lens' PutPermissionResponse Text

A universally unique identifier (UUID) for the revision of the resource-based policy that includes the added permissions. The JSON-formatted policy is in the policy element of the response.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The Amazon Resource Name (ARN) of the resource that contains the tags to return.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe (HashMap Text Text))

The list of tags assigned to the specified resource. This is the list of tags returned in the response.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

RemovePermission

removePermission_actionGroup :: Lens' RemovePermission ActionGroup

Specifies an action group that contains the permissions to remove from a profiling group's resource-based policy. One action group is supported, agentPermissions, which grants ConfigureAgent and PostAgentProfile permissions.

removePermission_profilingGroupName :: Lens' RemovePermission Text

The name of the profiling group.

removePermission_revisionId :: Lens' RemovePermission Text

A universally unique identifier (UUID) for the revision of the resource-based policy from which you want to remove permissions.

removePermissionResponse_httpStatus :: Lens' RemovePermissionResponse Int

The response's http status code.

removePermissionResponse_policy :: Lens' RemovePermissionResponse Text

The JSON-formatted resource-based policy on the profiling group after the specified permissions were removed.

removePermissionResponse_revisionId :: Lens' RemovePermissionResponse Text

A universally unique identifier (UUID) for the revision of the resource-based policy after the specified permissions were removed. The updated JSON-formatted policy is in the policy element of the response.

SubmitFeedback

submitFeedback_comment :: Lens' SubmitFeedback (Maybe Text)

Optional feedback about this anomaly.

submitFeedback_anomalyInstanceId :: Lens' SubmitFeedback Text

The universally unique identifier (UUID) of the AnomalyInstance object that is included in the analysis data.

submitFeedback_profilingGroupName :: Lens' SubmitFeedback Text

The name of the profiling group that is associated with the analysis data.

submitFeedback_type :: Lens' SubmitFeedback FeedbackType

The feedback tpye. Thee are two valid values, Positive and Negative.

submitFeedbackResponse_httpStatus :: Lens' SubmitFeedbackResponse Int

The response's http status code.

CreateProfilingGroup

createProfilingGroup_computePlatform :: Lens' CreateProfilingGroup (Maybe ComputePlatform)

The compute platform of the profiling group. Use AWSLambda if your application runs on AWS Lambda. Use Default if your application runs on a compute platform that is not AWS Lambda, such an Amazon EC2 instance, an on-premises server, or a different platform. If not specified, Default is used.

createProfilingGroup_agentOrchestrationConfig :: Lens' CreateProfilingGroup (Maybe AgentOrchestrationConfig)

Specifies whether profiling is enabled or disabled for the created profiling group.

createProfilingGroup_tags :: Lens' CreateProfilingGroup (Maybe (HashMap Text Text))

A list of tags to add to the created profiling group.

createProfilingGroup_clientToken :: Lens' CreateProfilingGroup Text

Amazon CodeGuru Profiler uses this universally unique identifier (UUID) to prevent the accidental creation of duplicate profiling groups if there are failures and retries.

createProfilingGroup_profilingGroupName :: Lens' CreateProfilingGroup Text

The name of the profiling group to create.

createProfilingGroupResponse_httpStatus :: Lens' CreateProfilingGroupResponse Int

The response's http status code.

createProfilingGroupResponse_profilingGroup :: Lens' CreateProfilingGroupResponse ProfilingGroupDescription

The returned ProfilingGroupDescription object that contains information about the created profiling group.

RemoveNotificationChannel

removeNotificationChannel_channelId :: Lens' RemoveNotificationChannel Text

The id of the channel that we want to stop receiving notifications.

removeNotificationChannel_profilingGroupName :: Lens' RemoveNotificationChannel Text

The name of the profiling group we want to change notification configuration for.

removeNotificationChannelResponse_notificationConfiguration :: Lens' RemoveNotificationChannelResponse (Maybe NotificationConfiguration)

The new notification configuration for this profiling group.

removeNotificationChannelResponse_httpStatus :: Lens' RemoveNotificationChannelResponse Int

The response's http status code.

UpdateProfilingGroup

updateProfilingGroup_agentOrchestrationConfig :: Lens' UpdateProfilingGroup AgentOrchestrationConfig

Specifies whether profiling is enabled or disabled for a profiling group.

updateProfilingGroup_profilingGroupName :: Lens' UpdateProfilingGroup Text

The name of the profiling group to update.

updateProfilingGroupResponse_httpStatus :: Lens' UpdateProfilingGroupResponse Int

The response's http status code.

updateProfilingGroupResponse_profilingGroup :: Lens' UpdateProfilingGroupResponse ProfilingGroupDescription

A ProfilingGroupDescription that contains information about the returned updated profiling group.

DeleteProfilingGroup

deleteProfilingGroup_profilingGroupName :: Lens' DeleteProfilingGroup Text

The name of the profiling group to delete.

deleteProfilingGroupResponse_httpStatus :: Lens' DeleteProfilingGroupResponse Int

The response's http status code.

ListFindingsReports

listFindingsReports_nextToken :: Lens' ListFindingsReports (Maybe Text)

The nextToken value returned from a previous paginated ListFindingsReportsRequest request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value.

This token should be treated as an opaque identifier that is only used to retrieve the next items in a list and not for other programmatic purposes.

listFindingsReports_dailyReportsOnly :: Lens' ListFindingsReports (Maybe Bool)

A Boolean value indicating whether to only return reports from daily profiles. If set to True, only analysis data from daily profiles is returned. If set to False, analysis data is returned from smaller time windows (for example, one hour).

listFindingsReports_maxResults :: Lens' ListFindingsReports (Maybe Natural)

The maximum number of report results returned by ListFindingsReports in paginated output. When this parameter is used, ListFindingsReports only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListFindingsReports request with the returned nextToken value.

listFindingsReports_endTime :: Lens' ListFindingsReports UTCTime

The end time of the profile to get analysis data about. You must specify startTime and endTime. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

listFindingsReports_profilingGroupName :: Lens' ListFindingsReports Text

The name of the profiling group from which to search for analysis data.

listFindingsReports_startTime :: Lens' ListFindingsReports UTCTime

The start time of the profile to get analysis data about. You must specify startTime and endTime. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

listFindingsReportsResponse_nextToken :: Lens' ListFindingsReportsResponse (Maybe Text)

The nextToken value to include in a future ListFindingsReports request. When the results of a ListFindingsReports request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

listFindingsReportsResponse_httpStatus :: Lens' ListFindingsReportsResponse Int

The response's http status code.

listFindingsReportsResponse_findingsReportSummaries :: Lens' ListFindingsReportsResponse [FindingsReportSummary]

The list of analysis results summaries.

ListProfileTimes

listProfileTimes_orderBy :: Lens' ListProfileTimes (Maybe OrderBy)

The order (ascending or descending by start time of the profile) to use when listing profiles. Defaults to TIMESTAMP_DESCENDING.

listProfileTimes_nextToken :: Lens' ListProfileTimes (Maybe Text)

The nextToken value returned from a previous paginated ListProfileTimes request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value.

This token should be treated as an opaque identifier that is only used to retrieve the next items in a list and not for other programmatic purposes.

listProfileTimes_maxResults :: Lens' ListProfileTimes (Maybe Natural)

The maximum number of profile time results returned by ListProfileTimes in paginated output. When this parameter is used, ListProfileTimes only returns maxResults results in a single page with a nextToken response element. The remaining results of the initial request can be seen by sending another ListProfileTimes request with the returned nextToken value.

listProfileTimes_endTime :: Lens' ListProfileTimes UTCTime

The end time of the time range from which to list the profiles.

listProfileTimes_period :: Lens' ListProfileTimes AggregationPeriod

The aggregation period. This specifies the period during which an aggregation profile collects posted agent profiles for a profiling group. There are 3 valid values.

P1D — 1 day

PT1H — 1 hour

PT5M — 5 minutes

listProfileTimes_profilingGroupName :: Lens' ListProfileTimes Text

The name of the profiling group.

listProfileTimes_startTime :: Lens' ListProfileTimes UTCTime

The start time of the time range from which to list the profiles.

listProfileTimesResponse_nextToken :: Lens' ListProfileTimesResponse (Maybe Text)

The nextToken value to include in a future ListProfileTimes request. When the results of a ListProfileTimes request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

listProfileTimesResponse_httpStatus :: Lens' ListProfileTimesResponse Int

The response's http status code.

listProfileTimesResponse_profileTimes :: Lens' ListProfileTimesResponse [ProfileTime]

The list of start times of the available profiles for the aggregation period in the specified time range.

PostAgentProfile

postAgentProfile_profileToken :: Lens' PostAgentProfile (Maybe Text)

Amazon CodeGuru Profiler uses this universally unique identifier (UUID) to prevent the accidental submission of duplicate profiling data if there are failures and retries.

postAgentProfile_agentProfile :: Lens' PostAgentProfile ByteString

The submitted profiling data.

postAgentProfile_contentType :: Lens' PostAgentProfile Text

The format of the submitted profiling data. The format maps to the Accept and Content-Type headers of the HTTP request. You can specify one of the following: or the default .

 <ul> <li> <p> <code>application/json</code> — standard JSON format </p> </li> <li> <p> <code>application/x-amzn-ion</code> — the Amazon Ion data format. For more information, see <a href="http://amzn.github.io/ion-docs/">Amazon Ion</a>. </p> </li> </ul>

postAgentProfile_profilingGroupName :: Lens' PostAgentProfile Text

The name of the profiling group with the aggregated profile that receives the submitted profiling data.

postAgentProfileResponse_httpStatus :: Lens' PostAgentProfileResponse Int

The response's http status code.

GetProfile

getProfile_startTime :: Lens' GetProfile (Maybe UTCTime)

The start time of the profile to get. Specify using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

 <p> If you specify <code>startTime</code>, then you must also specify <code>period</code> or <code>endTime</code>, but not both. </p>

getProfile_period :: Lens' GetProfile (Maybe Text)

Used with startTime or endTime to specify the time range for the returned aggregated profile. Specify using the ISO 8601 format. For example, P1DT1H1M1S.

 <p> To get the latest aggregated profile, specify only <code>period</code>. </p>

getProfile_accept :: Lens' GetProfile (Maybe Text)

The format of the returned profiling data. The format maps to the Accept and Content-Type headers of the HTTP request. You can specify one of the following: or the default .

 <ul> <li> <p> <code>application/json</code> — standard JSON format </p> </li> <li> <p> <code>application/x-amzn-ion</code> — the Amazon Ion data format. For more information, see <a href="http://amzn.github.io/ion-docs/">Amazon Ion</a>. </p> </li> </ul>

getProfile_endTime :: Lens' GetProfile (Maybe UTCTime)

The end time of the requested profile. Specify using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

If you specify endTime, then you must also specify period or startTime, but not both.

getProfile_maxDepth :: Lens' GetProfile (Maybe Natural)

The maximum depth of the stacks in the code that is represented in the aggregated profile. For example, if CodeGuru Profiler finds a method A, which calls method B, which calls method C, which calls method D, then the depth is 4. If the maxDepth is set to 2, then the aggregated profile contains representations of methods A and B.

getProfile_profilingGroupName :: Lens' GetProfile Text

The name of the profiling group to get.

getProfileResponse_contentEncoding :: Lens' GetProfileResponse (Maybe Text)

The content encoding of the profile.

getProfileResponse_httpStatus :: Lens' GetProfileResponse Int

The response's http status code.

getProfileResponse_contentType :: Lens' GetProfileResponse Text

The content type of the profile in the payload. It is either application/json or the default application/x-amzn-ion.

getProfileResponse_profile :: Lens' GetProfileResponse ByteString

Information about the profile.

ListProfilingGroups

listProfilingGroups_includeDescription :: Lens' ListProfilingGroups (Maybe Bool)

A Boolean value indicating whether to include a description. If true, then a list of ProfilingGroupDescription objects that contain detailed information about profiling groups is returned. If false, then a list of profiling group names is returned.

listProfilingGroups_nextToken :: Lens' ListProfilingGroups (Maybe Text)

The nextToken value returned from a previous paginated ListProfilingGroups request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value.

This token should be treated as an opaque identifier that is only used to retrieve the next items in a list and not for other programmatic purposes.

listProfilingGroups_maxResults :: Lens' ListProfilingGroups (Maybe Natural)

The maximum number of profiling groups results returned by ListProfilingGroups in paginated output. When this parameter is used, ListProfilingGroups only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListProfilingGroups request with the returned nextToken value.

listProfilingGroupsResponse_nextToken :: Lens' ListProfilingGroupsResponse (Maybe Text)

The nextToken value to include in a future ListProfilingGroups request. When the results of a ListProfilingGroups request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

listProfilingGroupsResponse_profilingGroups :: Lens' ListProfilingGroupsResponse (Maybe [ProfilingGroupDescription])

A returned list ProfilingGroupDescription objects. A list of ProfilingGroupDescription objects is returned only if includeDescription is true, otherwise a list of profiling group names is returned.

listProfilingGroupsResponse_httpStatus :: Lens' ListProfilingGroupsResponse Int

The response's http status code.

listProfilingGroupsResponse_profilingGroupNames :: Lens' ListProfilingGroupsResponse [Text]

A returned list of profiling group names. A list of the names is returned only if includeDescription is false, otherwise a list of ProfilingGroupDescription objects is returned.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The Amazon Resource Name (ARN) of the resource that the tags are added to.

tagResource_tags :: Lens' TagResource (HashMap Text Text)

The list of tags that are added to the specified resource.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

GetNotificationConfiguration

getNotificationConfiguration_profilingGroupName :: Lens' GetNotificationConfiguration Text

The name of the profiling group we want to get the notification configuration for.

getNotificationConfigurationResponse_httpStatus :: Lens' GetNotificationConfigurationResponse Int

The response's http status code.

getNotificationConfigurationResponse_notificationConfiguration :: Lens' GetNotificationConfigurationResponse NotificationConfiguration

The current notification configuration for this profiling group.

BatchGetFrameMetricData

batchGetFrameMetricData_targetResolution :: Lens' BatchGetFrameMetricData (Maybe AggregationPeriod)

The requested resolution of time steps for the returned time series of values. If the requested target resolution is not available due to data not being retained we provide a best effort result by falling back to the most granular available resolution after the target resolution. There are 3 valid values.

P1D — 1 day

PT1H — 1 hour

PT5M — 5 minutes

batchGetFrameMetricData_frameMetrics :: Lens' BatchGetFrameMetricData (Maybe [FrameMetric])

The details of the metrics that are used to request a time series of values. The metric includes the name of the frame, the aggregation type to calculate the metric value for the frame, and the thread states to use to get the count for the metric value of the frame.

batchGetFrameMetricData_startTime :: Lens' BatchGetFrameMetricData (Maybe UTCTime)

The start time of the time period for the frame metrics used to return the time series values. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

batchGetFrameMetricData_period :: Lens' BatchGetFrameMetricData (Maybe Text)

The duration of the frame metrics used to return the time series values. Specify using the ISO 8601 format. The maximum period duration is one day (PT24H or P1D).

batchGetFrameMetricData_endTime :: Lens' BatchGetFrameMetricData (Maybe UTCTime)

The end time of the time period for the returned time series values. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

batchGetFrameMetricData_profilingGroupName :: Lens' BatchGetFrameMetricData Text

The name of the profiling group associated with the the frame metrics used to return the time series values.

batchGetFrameMetricDataResponse_httpStatus :: Lens' BatchGetFrameMetricDataResponse Int

The response's http status code.

batchGetFrameMetricDataResponse_endTime :: Lens' BatchGetFrameMetricDataResponse UTCTime

The end time of the time period for the returned time series values. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

batchGetFrameMetricDataResponse_endTimes :: Lens' BatchGetFrameMetricDataResponse [TimestampStructure]

List of instances, or time steps, in the time series. For example, if the period is one day (PT24H)), and the resolution is five minutes (PT5M), then there are 288 endTimes in the list that are each five minutes appart.

batchGetFrameMetricDataResponse_frameMetricData :: Lens' BatchGetFrameMetricDataResponse [FrameMetricDatum]

Details of the metrics to request a time series of values. The metric includes the name of the frame, the aggregation type to calculate the metric value for the frame, and the thread states to use to get the count for the metric value of the frame.

batchGetFrameMetricDataResponse_resolution :: Lens' BatchGetFrameMetricDataResponse AggregationPeriod

Resolution or granularity of the profile data used to generate the time series. This is the value used to jump through time steps in a time series. There are 3 valid values.

P1D — 1 day

PT1H — 1 hour

PT5M — 5 minutes

batchGetFrameMetricDataResponse_startTime :: Lens' BatchGetFrameMetricDataResponse UTCTime

The start time of the time period for the returned time series values. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

batchGetFrameMetricDataResponse_unprocessedEndTimes :: Lens' BatchGetFrameMetricDataResponse (HashMap Text [TimestampStructure])

List of instances which remained unprocessed. This will create a missing time step in the list of end times.

UntagResource

untagResource_resourceArn :: Lens' UntagResource Text

The Amazon Resource Name (ARN) of the resource that contains the tags to remove.

untagResource_tagKeys :: Lens' UntagResource [Text]

A list of tag keys. Existing tags of resources with keys in this list are removed from the specified resource.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

GetFindingsReportAccountSummary

getFindingsReportAccountSummary_nextToken :: Lens' GetFindingsReportAccountSummary (Maybe Text)

The nextToken value returned from a previous paginated GetFindingsReportAccountSummary request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value.

This token should be treated as an opaque identifier that is only used to retrieve the next items in a list and not for other programmatic purposes.

getFindingsReportAccountSummary_dailyReportsOnly :: Lens' GetFindingsReportAccountSummary (Maybe Bool)

A Boolean value indicating whether to only return reports from daily profiles. If set to True, only analysis data from daily profiles is returned. If set to False, analysis data is returned from smaller time windows (for example, one hour).

getFindingsReportAccountSummary_maxResults :: Lens' GetFindingsReportAccountSummary (Maybe Natural)

The maximum number of results returned by GetFindingsReportAccountSummary in paginated output. When this parameter is used, GetFindingsReportAccountSummary only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another GetFindingsReportAccountSummary request with the returned nextToken value.

getFindingsReportAccountSummaryResponse_nextToken :: Lens' GetFindingsReportAccountSummaryResponse (Maybe Text)

The nextToken value to include in a future GetFindingsReportAccountSummary request. When the results of a GetFindingsReportAccountSummary request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

getFindingsReportAccountSummaryResponse_httpStatus :: Lens' GetFindingsReportAccountSummaryResponse Int

The response's http status code.

getFindingsReportAccountSummaryResponse_reportSummaries :: Lens' GetFindingsReportAccountSummaryResponse [FindingsReportSummary]

The return list of FindingsReportSummary objects taht contain summaries of analysis results for all profiling groups in your AWS account.

GetPolicy

getPolicy_profilingGroupName :: Lens' GetPolicy Text

The name of the profiling group.

getPolicyResponse_httpStatus :: Lens' GetPolicyResponse Int

The response's http status code.

getPolicyResponse_policy :: Lens' GetPolicyResponse Text

The JSON-formatted resource-based policy attached to the ProfilingGroup.

getPolicyResponse_revisionId :: Lens' GetPolicyResponse Text

A unique identifier for the current revision of the returned policy.

ConfigureAgent

configureAgent_fleetInstanceId :: Lens' ConfigureAgent (Maybe Text)

A universally unique identifier (UUID) for a profiling instance. For example, if the profiling instance is an Amazon EC2 instance, it is the instance ID. If it is an AWS Fargate container, it is the container's task ID.

configureAgent_metadata :: Lens' ConfigureAgent (Maybe (HashMap MetadataField Text))

Metadata captured about the compute platform the agent is running on. It includes information about sampling and reporting. The valid fields are:

COMPUTE_PLATFORM - The compute platform on which the agent is running

AGENT_ID - The ID for an agent instance.

AWS_REQUEST_ID - The AWS request ID of a Lambda invocation.

EXECUTION_ENVIRONMENT - The execution environment a Lambda function is running on.

LAMBDA_FUNCTION_ARN - The Amazon Resource Name (ARN) that is used to invoke a Lambda function.

LAMBDA_MEMORY_LIMIT_IN_MB - The memory allocated to a Lambda function.

LAMBDA_REMAINING_TIME_IN_MILLISECONDS - The time in milliseconds before execution of a Lambda function times out.

LAMBDA_TIME_GAP_BETWEEN_INVOKES_IN_MILLISECONDS - The time in milliseconds between two invocations of a Lambda function.

LAMBDA_PREVIOUS_EXECUTION_TIME_IN_MILLISECONDS - The time in milliseconds for the previous Lambda invocation.

configureAgent_profilingGroupName :: Lens' ConfigureAgent Text

The name of the profiling group for which the configured agent is collecting profiling data.

configureAgentResponse_httpStatus :: Lens' ConfigureAgentResponse Int

The response's http status code.

configureAgentResponse_configuration :: Lens' ConfigureAgentResponse AgentConfiguration

An AgentConfiguration object that specifies if an agent profiles or not and for how long to return profiling data.

Types

AgentConfiguration

agentConfiguration_agentParameters :: Lens' AgentConfiguration (Maybe (HashMap AgentParameterField Text))

Parameters used by the profiler. The valid parameters are:

MaxStackDepth - The maximum depth of the stacks in the code that is represented in the profile. For example, if CodeGuru Profiler finds a method A, which calls method B, which calls method C, which calls method D, then the depth is 4. If the maxDepth is set to 2, then the profiler evaluates A and B.

MemoryUsageLimitPercent - The percentage of memory that is used by the profiler.

MinimumTimeForReportingInMilliseconds - The minimum time in milliseconds between sending reports.

ReportingIntervalInMilliseconds - The reporting interval in milliseconds used to report profiles.

SamplingIntervalInMilliseconds - The sampling interval in milliseconds that is used to profile samples.

agentConfiguration_periodInSeconds :: Lens' AgentConfiguration Int

How long a profiling agent should send profiling data using ConfigureAgent . For example, if this is set to 300, the profiling agent calls ConfigureAgent every 5 minutes to submit the profiled data collected during that period.

agentConfiguration_shouldProfile :: Lens' AgentConfiguration Bool

A Boolean that specifies whether the profiling agent collects profiling data or not. Set to true to enable profiling.

AgentOrchestrationConfig

agentOrchestrationConfig_profilingEnabled :: Lens' AgentOrchestrationConfig Bool

A Boolean that specifies whether the profiling agent collects profiling data or not. Set to true to enable profiling.

AggregatedProfileTime

aggregatedProfileTime_period :: Lens' AggregatedProfileTime (Maybe AggregationPeriod)

The aggregation period. This indicates the period during which an aggregation profile collects posted agent profiles for a profiling group. Use one of three valid durations that are specified using the ISO 8601 format.

P1D — 1 day

PT1H — 1 hour

PT5M — 5 minutes

aggregatedProfileTime_start :: Lens' AggregatedProfileTime (Maybe UTCTime)

The time that aggregation of posted agent profiles for a profiling group starts. The aggregation profile contains profiles posted by the agent starting at this time for an aggregation period specified by the period property of the AggregatedProfileTime object.

Specify start using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

Anomaly

anomaly_instances :: Lens' Anomaly [AnomalyInstance]

A list of the instances of the detected anomalies during the requested period.

anomaly_metric :: Lens' Anomaly Metric

Details about the metric that the analysis used when it detected the anomaly. The metric includes the name of the frame that was analyzed with the type and thread states used to derive the metric value for that frame.

anomaly_reason :: Lens' Anomaly Text

The reason for which metric was flagged as anomalous.

AnomalyInstance

anomalyInstance_endTime :: Lens' AnomalyInstance (Maybe UTCTime)

The end time of the period during which the metric is flagged as anomalous. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

anomalyInstance_userFeedback :: Lens' AnomalyInstance (Maybe UserFeedback)

Feedback type on a specific instance of anomaly submitted by the user.

anomalyInstance_id :: Lens' AnomalyInstance Text

The universally unique identifier (UUID) of an instance of an anomaly in a metric.

anomalyInstance_startTime :: Lens' AnomalyInstance UTCTime

The start time of the period during which the metric is flagged as anomalous. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

Channel

channel_id :: Lens' Channel (Maybe Text)

Unique identifier for each Channel in the notification configuration of a Profiling Group. A random UUID for channelId is used when adding a channel to the notification configuration if not specified in the request.

channel_eventPublishers :: Lens' Channel (NonEmpty EventPublisher)

List of publishers for different type of events that may be detected in an application from the profile. Anomaly detection is the only event publisher in Profiler.

channel_uri :: Lens' Channel Text

Unique arn of the resource to be used for notifications. We support a valid SNS topic arn as a channel uri.

FindingsReportSummary

findingsReportSummary_profileStartTime :: Lens' FindingsReportSummary (Maybe UTCTime)

The start time of the profile the analysis data is about. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

findingsReportSummary_profileEndTime :: Lens' FindingsReportSummary (Maybe UTCTime)

The end time of the period during which the metric is flagged as anomalous. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

findingsReportSummary_id :: Lens' FindingsReportSummary (Maybe Text)

The universally unique identifier (UUID) of the recommendation report.

findingsReportSummary_profilingGroupName :: Lens' FindingsReportSummary (Maybe Text)

The name of the profiling group that is associated with the analysis data.

findingsReportSummary_totalNumberOfFindings :: Lens' FindingsReportSummary (Maybe Int)

The total number of different recommendations that were found by the analysis.

FrameMetric

frameMetric_frameName :: Lens' FrameMetric Text

Name of the method common across the multiple occurrences of a frame in an application profile.

frameMetric_threadStates :: Lens' FrameMetric [Text]

List of application runtime thread states used to get the counts for a frame a derive a metric value.

frameMetric_type :: Lens' FrameMetric MetricType

A type of aggregation that specifies how a metric for a frame is analyzed. The supported value AggregatedRelativeTotalTime is an aggregation of the metric value for one frame that is calculated across the occurrences of all frames in a profile.

FrameMetricDatum

frameMetricDatum_frameMetric :: Lens' FrameMetricDatum FrameMetric

Undocumented member.

frameMetricDatum_values :: Lens' FrameMetricDatum [Double]

A list of values that are associated with a frame metric.

Match

match_thresholdBreachValue :: Lens' Match (Maybe Double)

The value in the profile data that exceeded the recommendation threshold.

match_frameAddress :: Lens' Match (Maybe Text)

The location in the profiling graph that contains a recommendation found during analysis.

match_targetFramesIndex :: Lens' Match (Maybe Int)

The target frame that triggered a match.

Metric

metric_frameName :: Lens' Metric Text

The name of the method that appears as a frame in any stack in a profile.

metric_threadStates :: Lens' Metric [Text]

The list of application runtime thread states that is used to calculate the metric value for the frame.

metric_type :: Lens' Metric MetricType

A type that specifies how a metric for a frame is analyzed. The supported value AggregatedRelativeTotalTime is an aggregation of the metric value for one frame that is calculated across the occurences of all frames in a profile.

NotificationConfiguration

notificationConfiguration_channels :: Lens' NotificationConfiguration (Maybe (NonEmpty Channel))

List of up to two channels to be used for sending notifications for events detected from the application profile.

Pattern

pattern_thresholdPercent :: Lens' Pattern (Maybe Double)

The percentage of time an application spends in one method that triggers a recommendation. The percentage of time is the same as the percentage of the total gathered sample counts during analysis.

pattern_targetFrames :: Lens' Pattern (Maybe [[Text]])

A list of frame names that were searched during the analysis that generated a recommendation.

pattern_countersToAggregate :: Lens' Pattern (Maybe [Text])

A list of the different counters used to determine if there is a match.

pattern_name :: Lens' Pattern (Maybe Text)

The name for this pattern.

pattern_resolutionSteps :: Lens' Pattern (Maybe Text)

A string that contains the steps recommended to address the potential inefficiency.

pattern_id :: Lens' Pattern (Maybe Text)

The universally unique identifier (UUID) of this pattern.

pattern_description :: Lens' Pattern (Maybe Text)

The description of the recommendation. This explains a potential inefficiency in a profiled application.

ProfileTime

profileTime_start :: Lens' ProfileTime (Maybe UTCTime)

The start time of a profile. It is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

ProfilingGroupDescription

profilingGroupDescription_computePlatform :: Lens' ProfilingGroupDescription (Maybe ComputePlatform)

The compute platform of the profiling group. If it is set to AWSLambda, then the profiled application runs on AWS Lambda. If it is set to Default, then the profiled application runs on a compute platform that is not AWS Lambda, such an Amazon EC2 instance, an on-premises server, or a different platform. The default is Default.

profilingGroupDescription_arn :: Lens' ProfilingGroupDescription (Maybe Text)

The Amazon Resource Name (ARN) identifying the profiling group resource.

profilingGroupDescription_createdAt :: Lens' ProfilingGroupDescription (Maybe UTCTime)

The time when the profiling group was created. Specify using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

profilingGroupDescription_name :: Lens' ProfilingGroupDescription (Maybe Text)

The name of the profiling group.

profilingGroupDescription_profilingStatus :: Lens' ProfilingGroupDescription (Maybe ProfilingStatus)

A ProfilingStatus object that includes information about the last time a profile agent pinged back, the last time a profile was received, and the aggregation period and start time for the most recent aggregated profile.

profilingGroupDescription_updatedAt :: Lens' ProfilingGroupDescription (Maybe UTCTime)

The date and time when the profiling group was last updated. Specify using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

profilingGroupDescription_agentOrchestrationConfig :: Lens' ProfilingGroupDescription (Maybe AgentOrchestrationConfig)

An AgentOrchestrationConfig object that indicates if the profiling group is enabled for profiled or not.

profilingGroupDescription_tags :: Lens' ProfilingGroupDescription (Maybe (HashMap Text Text))

A list of the tags that belong to this profiling group.

ProfilingStatus

profilingStatus_latestAgentOrchestratedAt :: Lens' ProfilingStatus (Maybe UTCTime)

The date and time when the profiling agent most recently pinged back. Specify using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

profilingStatus_latestAgentProfileReportedAt :: Lens' ProfilingStatus (Maybe UTCTime)

The date and time when the most recent profile was received. Specify using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

profilingStatus_latestAggregatedProfile :: Lens' ProfilingStatus (Maybe AggregatedProfileTime)

An AggregatedProfileTime object that contains the aggregation period and start time for an aggregated profile.

Recommendation

recommendation_allMatchesCount :: Lens' Recommendation Int

How many different places in the profile graph triggered a match.

recommendation_allMatchesSum :: Lens' Recommendation Double

How much of the total sample count is potentially affected.

recommendation_endTime :: Lens' Recommendation UTCTime

End time of the profile that was used by this analysis. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

recommendation_pattern :: Lens' Recommendation Pattern

The pattern that analysis recognized in the profile to make this recommendation.

recommendation_startTime :: Lens' Recommendation UTCTime

The start time of the profile that was used by this analysis. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

recommendation_topMatches :: Lens' Recommendation [Match]

List of the matches with most impact.

TimestampStructure

timestampStructure_value :: Lens' TimestampStructure UTCTime

A Timestamp. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

UserFeedback

userFeedback_type :: Lens' UserFeedback FeedbackType

Optional Positive or Negative feedback submitted by the user about whether the recommendation is useful or not.