Network.AWS.CodeGuruProfiler.Types.FindingsReportSummary

Description

Documentation

data FindingsReportSummary

Information about potential recommendations that might be created from the analysis of profiling data.

See: newFindingsReportSummary smart constructor.

Constructors

FindingsReportSummary'

Fields

profileStartTime :: Maybe POSIX

The start time of the profile the analysis data is about. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

profileEndTime :: Maybe POSIX

The end time of the period during which the metric is flagged as anomalous. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

id :: Maybe Text

The universally unique identifier (UUID) of the recommendation report.

profilingGroupName :: Maybe Text

The name of the profiling group that is associated with the analysis data.

totalNumberOfFindings :: Maybe Int

The total number of different recommendations that were found by the analysis.

newFindingsReportSummary :: FindingsReportSummary

Create a value of FindingsReportSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:profileStartTime:FindingsReportSummary', findingsReportSummary_profileStartTime - The start time of the profile the analysis data is about. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

$sel:profileEndTime:FindingsReportSummary', findingsReportSummary_profileEndTime - The end time of the period during which the metric is flagged as anomalous. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

$sel:id:FindingsReportSummary', findingsReportSummary_id - The universally unique identifier (UUID) of the recommendation report.

$sel:profilingGroupName:FindingsReportSummary', findingsReportSummary_profilingGroupName - The name of the profiling group that is associated with the analysis data.

$sel:totalNumberOfFindings:FindingsReportSummary', findingsReportSummary_totalNumberOfFindings - The total number of different recommendations that were found by the analysis.

findingsReportSummary_profileStartTime :: Lens' FindingsReportSummary (Maybe UTCTime)

The start time of the profile the analysis data is about. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

findingsReportSummary_profileEndTime :: Lens' FindingsReportSummary (Maybe UTCTime)

The end time of the period during which the metric is flagged as anomalous. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

findingsReportSummary_id :: Lens' FindingsReportSummary (Maybe Text)

The universally unique identifier (UUID) of the recommendation report.

findingsReportSummary_profilingGroupName :: Lens' FindingsReportSummary (Maybe Text)

The name of the profiling group that is associated with the analysis data.

findingsReportSummary_totalNumberOfFindings :: Lens' FindingsReportSummary (Maybe Int)

The total number of different recommendations that were found by the analysis.