Network.AWS.CodeGuruProfiler.Types.ProfilingStatus

Description

Documentation

data ProfilingStatus

Profiling status includes information about the last time a profile agent pinged back, the last time a profile was received, and the aggregation period and start time for the most recent aggregated profile.

See: newProfilingStatus smart constructor.

Constructors

ProfilingStatus'

Fields

latestAgentOrchestratedAt :: Maybe POSIX

The date and time when the profiling agent most recently pinged back. Specify using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

latestAgentProfileReportedAt :: Maybe POSIX

The date and time when the most recent profile was received. Specify using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

latestAggregatedProfile :: Maybe AggregatedProfileTime

An AggregatedProfileTime object that contains the aggregation period and start time for an aggregated profile.

newProfilingStatus :: ProfilingStatus

Create a value of ProfilingStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:latestAgentOrchestratedAt:ProfilingStatus', profilingStatus_latestAgentOrchestratedAt - The date and time when the profiling agent most recently pinged back. Specify using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

$sel:latestAgentProfileReportedAt:ProfilingStatus', profilingStatus_latestAgentProfileReportedAt - The date and time when the most recent profile was received. Specify using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

$sel:latestAggregatedProfile:ProfilingStatus', profilingStatus_latestAggregatedProfile - An AggregatedProfileTime object that contains the aggregation period and start time for an aggregated profile.

profilingStatus_latestAgentOrchestratedAt :: Lens' ProfilingStatus (Maybe UTCTime)

The date and time when the profiling agent most recently pinged back. Specify using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

profilingStatus_latestAgentProfileReportedAt :: Lens' ProfilingStatus (Maybe UTCTime)

The date and time when the most recent profile was received. Specify using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

profilingStatus_latestAggregatedProfile :: Lens' ProfilingStatus (Maybe AggregatedProfileTime)

An AggregatedProfileTime object that contains the aggregation period and start time for an aggregated profile.