Network.AWS.CodeGuruReviewer.DescribeRepositoryAssociation

Description

Returns a RepositoryAssociation object that contains information about the requested repository association.

Creating a Request

data DescribeRepositoryAssociation

See: newDescribeRepositoryAssociation smart constructor.

Constructors

DescribeRepositoryAssociation'

Fields

associationArn :: Text

The Amazon Resource Name (ARN) of the RepositoryAssociation object. You can retrieve this ARN by calling ListRepositoryAssociations .

newDescribeRepositoryAssociation

Arguments

:: Text
$sel:associationArn:DescribeRepositoryAssociation'

-> DescribeRepositoryAssociation

Create a value of DescribeRepositoryAssociation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:associationArn:DescribeRepositoryAssociation', describeRepositoryAssociation_associationArn - The Amazon Resource Name (ARN) of the RepositoryAssociation object. You can retrieve this ARN by calling ListRepositoryAssociations .

Request Lenses

describeRepositoryAssociation_associationArn :: Lens' DescribeRepositoryAssociation Text

The Amazon Resource Name (ARN) of the RepositoryAssociation object. You can retrieve this ARN by calling ListRepositoryAssociations .

Destructuring the Response

data DescribeRepositoryAssociationResponse

See: newDescribeRepositoryAssociationResponse smart constructor.

Constructors

DescribeRepositoryAssociationResponse'

Fields

repositoryAssociation :: Maybe RepositoryAssociation

Information about the repository association.

tags :: Maybe (HashMap Text Text)

An array of key-value pairs used to tag an associated repository. A tag is a custom attribute label with two parts:

A tag key (for example, CostCenter, Environment, Project, or Secret). Tag keys are case sensitive.

An optional field known as a tag value (for example, 111122223333, Production, or a team name). Omitting the tag value is the same as using an empty string. Like tag keys, tag values are case sensitive.

httpStatus :: Int

The response's http status code.

newDescribeRepositoryAssociationResponse

Arguments

:: Int
$sel:httpStatus:DescribeRepositoryAssociationResponse'

-> DescribeRepositoryAssociationResponse

Create a value of DescribeRepositoryAssociationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:repositoryAssociation:DescribeRepositoryAssociationResponse', describeRepositoryAssociationResponse_repositoryAssociation - Information about the repository association.

$sel:tags:DescribeRepositoryAssociationResponse', describeRepositoryAssociationResponse_tags - An array of key-value pairs used to tag an associated repository. A tag is a custom attribute label with two parts:

A tag key (for example, CostCenter, Environment, Project, or Secret). Tag keys are case sensitive.

An optional field known as a tag value (for example, 111122223333, Production, or a team name). Omitting the tag value is the same as using an empty string. Like tag keys, tag values are case sensitive.

$sel:httpStatus:DescribeRepositoryAssociationResponse', describeRepositoryAssociationResponse_httpStatus - The response's http status code.

Response Lenses

describeRepositoryAssociationResponse_repositoryAssociation :: Lens' DescribeRepositoryAssociationResponse (Maybe RepositoryAssociation)

Information about the repository association.

describeRepositoryAssociationResponse_tags :: Lens' DescribeRepositoryAssociationResponse (Maybe (HashMap Text Text))

An array of key-value pairs used to tag an associated repository. A tag is a custom attribute label with two parts:

A tag key (for example, CostCenter, Environment, Project, or Secret). Tag keys are case sensitive.

An optional field known as a tag value (for example, 111122223333, Production, or a team name). Omitting the tag value is the same as using an empty string. Like tag keys, tag values are case sensitive.

describeRepositoryAssociationResponse_httpStatus :: Lens' DescribeRepositoryAssociationResponse Int

The response's http status code.