Network.AWS.CodeGuruReviewer.Lens

Description

Operations

ListRecommendationFeedback

listRecommendationFeedback_userIds :: Lens' ListRecommendationFeedback (Maybe (NonEmpty Text))

An Amazon Web Services user's account ID or Amazon Resource Name (ARN). Use this ID to query the recommendation feedback for a code review from that user.

The UserId is an IAM principal that can be specified as an Amazon Web Services account ID or an Amazon Resource Name (ARN). For more information, see Specifying a Principal in the Amazon Web Services Identity and Access Management User Guide.

listRecommendationFeedback_nextToken :: Lens' ListRecommendationFeedback (Maybe Text)

If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged.

listRecommendationFeedback_recommendationIds :: Lens' ListRecommendationFeedback (Maybe (NonEmpty Text))

Used to query the recommendation feedback for a given recommendation.

listRecommendationFeedback_maxResults :: Lens' ListRecommendationFeedback (Maybe Natural)

The maximum number of results that are returned per call. The default is 100.

listRecommendationFeedback_codeReviewArn :: Lens' ListRecommendationFeedback Text

The Amazon Resource Name (ARN) of the CodeReview object.

listRecommendationFeedbackResponse_nextToken :: Lens' ListRecommendationFeedbackResponse (Maybe Text)

If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged.

listRecommendationFeedbackResponse_recommendationFeedbackSummaries :: Lens' ListRecommendationFeedbackResponse (Maybe [RecommendationFeedbackSummary])

Recommendation feedback summaries corresponding to the code review ARN.

listRecommendationFeedbackResponse_httpStatus :: Lens' ListRecommendationFeedbackResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The Amazon Resource Name (ARN) of the RepositoryAssociation object. You can retrieve this ARN by calling ListRepositoryAssociations .

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe (HashMap Text Text))

An array of key-value pairs used to tag an associated repository. A tag is a custom attribute label with two parts:

A tag key (for example, CostCenter, Environment, Project, or Secret). Tag keys are case sensitive.

An optional field known as a tag value (for example, 111122223333, Production, or a team name). Omitting the tag value is the same as using an empty string. Like tag keys, tag values are case sensitive.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

DisassociateRepository

disassociateRepository_associationArn :: Lens' DisassociateRepository Text

The Amazon Resource Name (ARN) of the RepositoryAssociation object. You can retrieve this ARN by calling ListRepositoryAssociations .

disassociateRepositoryResponse_repositoryAssociation :: Lens' DisassociateRepositoryResponse (Maybe RepositoryAssociation)

Information about the disassociated repository.

disassociateRepositoryResponse_tags :: Lens' DisassociateRepositoryResponse (Maybe (HashMap Text Text))

An array of key-value pairs used to tag an associated repository. A tag is a custom attribute label with two parts:

A tag key (for example, CostCenter, Environment, Project, or Secret). Tag keys are case sensitive.

An optional field known as a tag value (for example, 111122223333, Production, or a team name). Omitting the tag value is the same as using an empty string. Like tag keys, tag values are case sensitive.

disassociateRepositoryResponse_httpStatus :: Lens' DisassociateRepositoryResponse Int

The response's http status code.

DescribeRepositoryAssociation

describeRepositoryAssociation_associationArn :: Lens' DescribeRepositoryAssociation Text

The Amazon Resource Name (ARN) of the RepositoryAssociation object. You can retrieve this ARN by calling ListRepositoryAssociations .

describeRepositoryAssociationResponse_repositoryAssociation :: Lens' DescribeRepositoryAssociationResponse (Maybe RepositoryAssociation)

Information about the repository association.

describeRepositoryAssociationResponse_tags :: Lens' DescribeRepositoryAssociationResponse (Maybe (HashMap Text Text))

An array of key-value pairs used to tag an associated repository. A tag is a custom attribute label with two parts:

A tag key (for example, CostCenter, Environment, Project, or Secret). Tag keys are case sensitive.

An optional field known as a tag value (for example, 111122223333, Production, or a team name). Omitting the tag value is the same as using an empty string. Like tag keys, tag values are case sensitive.

describeRepositoryAssociationResponse_httpStatus :: Lens' DescribeRepositoryAssociationResponse Int

The response's http status code.

DescribeCodeReview

describeCodeReview_codeReviewArn :: Lens' DescribeCodeReview Text

The Amazon Resource Name (ARN) of the CodeReview object.

describeCodeReviewResponse_codeReview :: Lens' DescribeCodeReviewResponse (Maybe CodeReview)

Information about the code review.

describeCodeReviewResponse_httpStatus :: Lens' DescribeCodeReviewResponse Int

The response's http status code.

ListRepositoryAssociations

listRepositoryAssociations_states :: Lens' ListRepositoryAssociations (Maybe (NonEmpty RepositoryAssociationState))

List of repository association states to use as a filter.

The valid repository association states are:

Associated: The repository association is complete.

Associating: CodeGuru Reviewer is:

Setting up pull request notifications. This is required for pull requests to trigger a CodeGuru Reviewer review.

If your repository ProviderType is GitHub, GitHub Enterprise Server, or Bitbucket, CodeGuru Reviewer creates webhooks in your repository to trigger CodeGuru Reviewer reviews. If you delete these webhooks, reviews of code in your repository cannot be triggered.

Setting up source code access. This is required for CodeGuru Reviewer to securely clone code in your repository.

Failed: The repository failed to associate or disassociate.

Disassociating: CodeGuru Reviewer is removing the repository's pull request notifications and source code access.

Disassociated: CodeGuru Reviewer successfully disassociated the repository. You can create a new association with this repository if you want to review source code in it later. You can control access to code reviews created in an associated repository with tags after it has been disassociated. For more information, see Using tags to control access to associated repositories in the Amazon CodeGuru Reviewer User Guide.

listRepositoryAssociations_owners :: Lens' ListRepositoryAssociations (Maybe (NonEmpty Text))

List of owners to use as a filter. For Amazon Web Services CodeCommit, it is the name of the CodeCommit account that was used to associate the repository. For other repository source providers, such as Bitbucket and GitHub Enterprise Server, this is name of the account that was used to associate the repository.

listRepositoryAssociations_providerTypes :: Lens' ListRepositoryAssociations (Maybe (NonEmpty ProviderType))

List of provider types to use as a filter.

listRepositoryAssociations_nextToken :: Lens' ListRepositoryAssociations (Maybe Text)

The nextToken value returned from a previous paginated ListRepositoryAssociations request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value.

Treat this token as an opaque identifier that is only used to retrieve the next items in a list and not for other programmatic purposes.

listRepositoryAssociations_names :: Lens' ListRepositoryAssociations (Maybe (NonEmpty Text))

List of repository names to use as a filter.

listRepositoryAssociations_maxResults :: Lens' ListRepositoryAssociations (Maybe Natural)

The maximum number of repository association results returned by ListRepositoryAssociations in paginated output. When this parameter is used, ListRepositoryAssociations only returns maxResults results in a single page with a nextToken response element. The remaining results of the initial request can be seen by sending another ListRepositoryAssociations request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, ListRepositoryAssociations returns up to 100 results and a nextToken value if applicable.

listRepositoryAssociationsResponse_nextToken :: Lens' ListRepositoryAssociationsResponse (Maybe Text)

The nextToken value to include in a future ListRecommendations request. When the results of a ListRecommendations request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

listRepositoryAssociationsResponse_repositoryAssociationSummaries :: Lens' ListRepositoryAssociationsResponse (Maybe [RepositoryAssociationSummary])

A list of repository associations that meet the criteria of the request.

listRepositoryAssociationsResponse_httpStatus :: Lens' ListRepositoryAssociationsResponse Int

The response's http status code.

DescribeRecommendationFeedback

describeRecommendationFeedback_userId :: Lens' DescribeRecommendationFeedback (Maybe Text)

Optional parameter to describe the feedback for a given user. If this is not supplied, it defaults to the user making the request.

The UserId is an IAM principal that can be specified as an Amazon Web Services account ID or an Amazon Resource Name (ARN). For more information, see Specifying a Principal in the Amazon Web Services Identity and Access Management User Guide.

describeRecommendationFeedback_codeReviewArn :: Lens' DescribeRecommendationFeedback Text

The Amazon Resource Name (ARN) of the CodeReview object.

describeRecommendationFeedback_recommendationId :: Lens' DescribeRecommendationFeedback Text

The recommendation ID that can be used to track the provided recommendations and then to collect the feedback.

describeRecommendationFeedbackResponse_recommendationFeedback :: Lens' DescribeRecommendationFeedbackResponse (Maybe RecommendationFeedback)

The recommendation feedback given by the user.

describeRecommendationFeedbackResponse_httpStatus :: Lens' DescribeRecommendationFeedbackResponse Int

The response's http status code.

ListRecommendations

listRecommendations_nextToken :: Lens' ListRecommendations (Maybe Text)

Pagination token.

listRecommendations_maxResults :: Lens' ListRecommendations (Maybe Natural)

The maximum number of results that are returned per call. The default is 100.

listRecommendations_codeReviewArn :: Lens' ListRecommendations Text

The Amazon Resource Name (ARN) of the CodeReview object.

listRecommendationsResponse_nextToken :: Lens' ListRecommendationsResponse (Maybe Text)

Pagination token.

listRecommendationsResponse_recommendationSummaries :: Lens' ListRecommendationsResponse (Maybe [RecommendationSummary])

List of recommendations for the requested code review.

listRecommendationsResponse_httpStatus :: Lens' ListRecommendationsResponse Int

The response's http status code.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The Amazon Resource Name (ARN) of the RepositoryAssociation object. You can retrieve this ARN by calling ListRepositoryAssociations .

tagResource_tags :: Lens' TagResource (HashMap Text Text)

An array of key-value pairs used to tag an associated repository. A tag is a custom attribute label with two parts:

A tag key (for example, CostCenter, Environment, Project, or Secret). Tag keys are case sensitive.

An optional field known as a tag value (for example, 111122223333, Production, or a team name). Omitting the tag value is the same as using an empty string. Like tag keys, tag values are case sensitive.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

CreateCodeReview

createCodeReview_clientRequestToken :: Lens' CreateCodeReview (Maybe Text)

Amazon CodeGuru Reviewer uses this value to prevent the accidental creation of duplicate code reviews if there are failures and retries.

createCodeReview_name :: Lens' CreateCodeReview Text

The name of the code review. The name of each code review in your Amazon Web Services account must be unique.

createCodeReview_repositoryAssociationArn :: Lens' CreateCodeReview Text

The Amazon Resource Name (ARN) of the RepositoryAssociation object. You can retrieve this ARN by calling ListRepositoryAssociations .

A code review can only be created on an associated repository. This is the ARN of the associated repository.

createCodeReview_type :: Lens' CreateCodeReview CodeReviewType

The type of code review to create. This is specified using a CodeReviewType object. You can create a code review only of type RepositoryAnalysis.

createCodeReviewResponse_codeReview :: Lens' CreateCodeReviewResponse (Maybe CodeReview)

Undocumented member.

createCodeReviewResponse_httpStatus :: Lens' CreateCodeReviewResponse Int

The response's http status code.

UntagResource

untagResource_resourceArn :: Lens' UntagResource Text

The Amazon Resource Name (ARN) of the RepositoryAssociation object. You can retrieve this ARN by calling ListRepositoryAssociations .

untagResource_tagKeys :: Lens' UntagResource (NonEmpty Text)

A list of the keys for each tag you want to remove from an associated repository.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

ListCodeReviews

listCodeReviews_states :: Lens' ListCodeReviews (Maybe (NonEmpty JobState))

List of states for filtering that needs to be applied before displaying the result. For example, states=[Pending] lists code reviews in the Pending state.

The valid code review states are:

Completed: The code review is complete.

Pending: The code review started and has not completed or failed.

Failed: The code review failed.

Deleting: The code review is being deleted.

listCodeReviews_providerTypes :: Lens' ListCodeReviews (Maybe (NonEmpty ProviderType))

List of provider types for filtering that needs to be applied before displaying the result. For example, providerTypes=[GitHub] lists code reviews from GitHub.

listCodeReviews_repositoryNames :: Lens' ListCodeReviews (Maybe (NonEmpty Text))

List of repository names for filtering that needs to be applied before displaying the result.

listCodeReviews_nextToken :: Lens' ListCodeReviews (Maybe Text)

If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged.

listCodeReviews_maxResults :: Lens' ListCodeReviews (Maybe Natural)

The maximum number of results that are returned per call. The default is 100.

listCodeReviews_type :: Lens' ListCodeReviews Type

The type of code reviews to list in the response.

listCodeReviewsResponse_codeReviewSummaries :: Lens' ListCodeReviewsResponse (Maybe [CodeReviewSummary])

A list of code reviews that meet the criteria of the request.

listCodeReviewsResponse_nextToken :: Lens' ListCodeReviewsResponse (Maybe Text)

Pagination token.

listCodeReviewsResponse_httpStatus :: Lens' ListCodeReviewsResponse Int

The response's http status code.

AssociateRepository

associateRepository_kmsKeyDetails :: Lens' AssociateRepository (Maybe KMSKeyDetails)

A KMSKeyDetails object that contains:

The encryption option for this repository association. It is either owned by Amazon Web Services Key Management Service (KMS) (AWS_OWNED_CMK) or customer managed (CUSTOMER_MANAGED_CMK).

The ID of the Amazon Web Services KMS key that is associated with this respository association.

associateRepository_clientRequestToken :: Lens' AssociateRepository (Maybe Text)

Amazon CodeGuru Reviewer uses this value to prevent the accidental creation of duplicate repository associations if there are failures and retries.

associateRepository_tags :: Lens' AssociateRepository (Maybe (HashMap Text Text))

An array of key-value pairs used to tag an associated repository. A tag is a custom attribute label with two parts:

A tag key (for example, CostCenter, Environment, Project, or Secret). Tag keys are case sensitive.

An optional field known as a tag value (for example, 111122223333, Production, or a team name). Omitting the tag value is the same as using an empty string. Like tag keys, tag values are case sensitive.

associateRepository_repository :: Lens' AssociateRepository Repository

The repository to associate.

associateRepositoryResponse_repositoryAssociation :: Lens' AssociateRepositoryResponse (Maybe RepositoryAssociation)

Information about the repository association.

associateRepositoryResponse_tags :: Lens' AssociateRepositoryResponse (Maybe (HashMap Text Text))

An array of key-value pairs used to tag an associated repository. A tag is a custom attribute label with two parts:

A tag key (for example, CostCenter, Environment, Project, or Secret). Tag keys are case sensitive.

An optional field known as a tag value (for example, 111122223333, Production, or a team name). Omitting the tag value is the same as using an empty string. Like tag keys, tag values are case sensitive.

associateRepositoryResponse_httpStatus :: Lens' AssociateRepositoryResponse Int

The response's http status code.

PutRecommendationFeedback

putRecommendationFeedback_codeReviewArn :: Lens' PutRecommendationFeedback Text

The Amazon Resource Name (ARN) of the CodeReview object.

putRecommendationFeedback_recommendationId :: Lens' PutRecommendationFeedback Text

The recommendation ID that can be used to track the provided recommendations and then to collect the feedback.

putRecommendationFeedback_reactions :: Lens' PutRecommendationFeedback [Reaction]

List for storing reactions. Reactions are utf-8 text code for emojis. If you send an empty list it clears all your feedback.

putRecommendationFeedbackResponse_httpStatus :: Lens' PutRecommendationFeedbackResponse Int

The response's http status code.

Types

BranchDiffSourceCodeType

branchDiffSourceCodeType_sourceBranchName :: Lens' BranchDiffSourceCodeType Text

The source branch for a diff in an associated repository.

branchDiffSourceCodeType_destinationBranchName :: Lens' BranchDiffSourceCodeType Text

The destination branch for a diff in an associated repository.

CodeArtifacts

codeArtifacts_buildArtifactsObjectKey :: Lens' CodeArtifacts (Maybe Text)

The S3 object key for a build artifacts .zip file that contains .jar or .class files. This is required for a code review with security analysis. For more information, see Create code reviews with security analysis in the Amazon CodeGuru Reviewer User Guide.

codeArtifacts_sourceCodeArtifactsObjectKey :: Lens' CodeArtifacts Text

The S3 object key for a source code .zip file. This is required for all code reviews.

CodeCommitRepository

codeCommitRepository_name :: Lens' CodeCommitRepository Text

The name of the Amazon Web Services CodeCommit repository. For more information, see repositoryName in the Amazon Web Services CodeCommit API Reference.

CodeReview

codeReview_associationArn :: Lens' CodeReview (Maybe Text)

The Amazon Resource Name (ARN) of the RepositoryAssociation that contains the reviewed source code. You can retrieve associated repository ARNs by calling ListRepositoryAssociations .

codeReview_state :: Lens' CodeReview (Maybe JobState)

The valid code review states are:

Completed: The code review is complete.

Pending: The code review started and has not completed or failed.

Failed: The code review failed.

Deleting: The code review is being deleted.

codeReview_metrics :: Lens' CodeReview (Maybe Metrics)

The statistics from the code review.

codeReview_pullRequestId :: Lens' CodeReview (Maybe Text)

The pull request ID for the code review.

codeReview_providerType :: Lens' CodeReview (Maybe ProviderType)

The type of repository that contains the reviewed code (for example, GitHub or Bitbucket).

codeReview_owner :: Lens' CodeReview (Maybe Text)

The owner of the repository. For an Amazon Web Services CodeCommit repository, this is the Amazon Web Services account ID of the account that owns the repository. For a GitHub, GitHub Enterprise Server, or Bitbucket repository, this is the username for the account that owns the repository. For an S3 repository, it can be the username or Amazon Web Services account ID.

codeReview_name :: Lens' CodeReview (Maybe Text)

The name of the code review.

codeReview_codeReviewArn :: Lens' CodeReview (Maybe Text)

The Amazon Resource Name (ARN) of the CodeReview object.

codeReview_repositoryName :: Lens' CodeReview (Maybe Text)

The name of the repository.

codeReview_type :: Lens' CodeReview (Maybe Type)

The type of code review.

codeReview_sourceCodeType :: Lens' CodeReview (Maybe SourceCodeType)

The type of the source code for the code review.

codeReview_stateReason :: Lens' CodeReview (Maybe Text)

The reason for the state of the code review.

codeReview_createdTimeStamp :: Lens' CodeReview (Maybe UTCTime)

The time, in milliseconds since the epoch, when the code review was created.

codeReview_analysisTypes :: Lens' CodeReview (Maybe [AnalysisType])

They types of analysis performed during a repository analysis or a pull request review. You can specify either Security, CodeQuality, or both.

codeReview_lastUpdatedTimeStamp :: Lens' CodeReview (Maybe UTCTime)

The time, in milliseconds since the epoch, when the code review was last updated.

CodeReviewSummary

codeReviewSummary_state :: Lens' CodeReviewSummary (Maybe JobState)

The state of the code review.

The valid code review states are:

Completed: The code review is complete.

Pending: The code review started and has not completed or failed.

Failed: The code review failed.

Deleting: The code review is being deleted.

codeReviewSummary_pullRequestId :: Lens' CodeReviewSummary (Maybe Text)

The pull request ID for the code review.

codeReviewSummary_providerType :: Lens' CodeReviewSummary (Maybe ProviderType)

The provider type of the repository association.

codeReviewSummary_owner :: Lens' CodeReviewSummary (Maybe Text)

The owner of the repository. For an Amazon Web Services CodeCommit repository, this is the Amazon Web Services account ID of the account that owns the repository. For a GitHub, GitHub Enterprise Server, or Bitbucket repository, this is the username for the account that owns the repository. For an S3 repository, it can be the username or Amazon Web Services account ID.

codeReviewSummary_name :: Lens' CodeReviewSummary (Maybe Text)

The name of the code review.

codeReviewSummary_codeReviewArn :: Lens' CodeReviewSummary (Maybe Text)

The Amazon Resource Name (ARN) of the CodeReview object.

codeReviewSummary_repositoryName :: Lens' CodeReviewSummary (Maybe Text)

The name of the repository.

codeReviewSummary_type :: Lens' CodeReviewSummary (Maybe Type)

The type of the code review.

codeReviewSummary_sourceCodeType :: Lens' CodeReviewSummary (Maybe SourceCodeType)

Undocumented member.

codeReviewSummary_metricsSummary :: Lens' CodeReviewSummary (Maybe MetricsSummary)

The statistics from the code review.

codeReviewSummary_createdTimeStamp :: Lens' CodeReviewSummary (Maybe UTCTime)

The time, in milliseconds since the epoch, when the code review was created.

codeReviewSummary_lastUpdatedTimeStamp :: Lens' CodeReviewSummary (Maybe UTCTime)

The time, in milliseconds since the epoch, when the code review was last updated.

CodeReviewType

codeReviewType_analysisTypes :: Lens' CodeReviewType (Maybe [AnalysisType])

They types of analysis performed during a repository analysis or a pull request review. You can specify either Security, CodeQuality, or both.

codeReviewType_repositoryAnalysis :: Lens' CodeReviewType RepositoryAnalysis

A code review that analyzes all code under a specified branch in an associated repository. The associated repository is specified using its ARN in CreateCodeReview .

CommitDiffSourceCodeType

commitDiffSourceCodeType_sourceCommit :: Lens' CommitDiffSourceCodeType (Maybe Text)

The SHA of the source commit used to generate a commit diff. This field is required for a pull request code review.

commitDiffSourceCodeType_mergeBaseCommit :: Lens' CommitDiffSourceCodeType (Maybe Text)

The SHA of the merge base of a commit.

commitDiffSourceCodeType_destinationCommit :: Lens' CommitDiffSourceCodeType (Maybe Text)

The SHA of the destination commit used to generate a commit diff. This field is required for a pull request code review.

EventInfo

eventInfo_state :: Lens' EventInfo (Maybe Text)

The state of an event. The state might be open, closed, or another state.

eventInfo_name :: Lens' EventInfo (Maybe Text)

The name of the event. The possible names are pull_request, workflow_dispatch, schedule, and push

KMSKeyDetails

kmsKeyDetails_encryptionOption :: Lens' KMSKeyDetails (Maybe EncryptionOption)

The encryption option for a repository association. It is either owned by Amazon Web Services Key Management Service (KMS) (AWS_OWNED_CMK) or customer managed (CUSTOMER_MANAGED_CMK).

kmsKeyDetails_kmsKeyId :: Lens' KMSKeyDetails (Maybe Text)

The ID of the Amazon Web Services KMS key that is associated with a respository association.

Metrics

metrics_findingsCount :: Lens' Metrics (Maybe Integer)

Total number of recommendations found in the code review.

metrics_meteredLinesOfCodeCount :: Lens' Metrics (Maybe Integer)

MeteredLinesOfCode is the number of lines of code in the repository where the code review happened. This does not include non-code lines such as comments and blank lines.

MetricsSummary

metricsSummary_findingsCount :: Lens' MetricsSummary (Maybe Integer)

Total number of recommendations found in the code review.

metricsSummary_meteredLinesOfCodeCount :: Lens' MetricsSummary (Maybe Integer)

Lines of code metered in the code review. For the initial code review pull request and all subsequent revisions, this includes all lines of code in the files added to the pull request. In subsequent revisions, for files that already existed in the pull request, this includes only the changed lines of code. In both cases, this does not include non-code lines such as comments and import statements. For example, if you submit a pull request containing 5 files, each with 500 lines of code, and in a subsequent revision you added a new file with 200 lines of code, and also modified a total of 25 lines across the initial 5 files, MeteredLinesOfCodeCount includes the first 5 files (5 * 500 = 2,500 lines), the new file (200 lines) and the 25 changed lines of code for a total of 2,725 lines of code.

RecommendationFeedback

recommendationFeedback_recommendationId :: Lens' RecommendationFeedback (Maybe Text)

The recommendation ID that can be used to track the provided recommendations. Later on it can be used to collect the feedback.

recommendationFeedback_userId :: Lens' RecommendationFeedback (Maybe Text)

The ID of the user that made the API call.

The UserId is an IAM principal that can be specified as an Amazon Web Services account ID or an Amazon Resource Name (ARN). For more information, see Specifying a Principal in the Amazon Web Services Identity and Access Management User Guide.

recommendationFeedback_reactions :: Lens' RecommendationFeedback (Maybe [Reaction])

List for storing reactions. Reactions are utf-8 text code for emojis. You can send an empty list to clear off all your feedback.

recommendationFeedback_codeReviewArn :: Lens' RecommendationFeedback (Maybe Text)

The Amazon Resource Name (ARN) of the CodeReview object.

recommendationFeedback_createdTimeStamp :: Lens' RecommendationFeedback (Maybe UTCTime)

The time at which the feedback was created.

recommendationFeedback_lastUpdatedTimeStamp :: Lens' RecommendationFeedback (Maybe UTCTime)

The time at which the feedback was last updated.

RecommendationFeedbackSummary

recommendationFeedbackSummary_recommendationId :: Lens' RecommendationFeedbackSummary (Maybe Text)

The recommendation ID that can be used to track the provided recommendations. Later on it can be used to collect the feedback.

recommendationFeedbackSummary_userId :: Lens' RecommendationFeedbackSummary (Maybe Text)

The ID of the user that gave the feedback.

The UserId is an IAM principal that can be specified as an Amazon Web Services account ID or an Amazon Resource Name (ARN). For more information, see Specifying a Principal in the Amazon Web Services Identity and Access Management User Guide.

recommendationFeedbackSummary_reactions :: Lens' RecommendationFeedbackSummary (Maybe [Reaction])

List for storing reactions. Reactions are utf-8 text code for emojis.

RecommendationSummary

recommendationSummary_recommendationId :: Lens' RecommendationSummary (Maybe Text)

The recommendation ID that can be used to track the provided recommendations. Later on it can be used to collect the feedback.

recommendationSummary_filePath :: Lens' RecommendationSummary (Maybe Text)

Name of the file on which a recommendation is provided.

recommendationSummary_severity :: Lens' RecommendationSummary (Maybe Severity)

The severity of the issue in the code that generated this recommendation.

recommendationSummary_ruleMetadata :: Lens' RecommendationSummary (Maybe RuleMetadata)

Metadata about a rule. Rule metadata includes an ID, a name, a list of tags, and a short and long description. CodeGuru Reviewer uses rules to analyze code. A rule's recommendation is included in analysis results if code is detected that violates the rule.

recommendationSummary_startLine :: Lens' RecommendationSummary (Maybe Int)

Start line from where the recommendation is applicable in the source commit or source branch.

recommendationSummary_endLine :: Lens' RecommendationSummary (Maybe Int)

Last line where the recommendation is applicable in the source commit or source branch. For a single line comment the start line and end line values are the same.

recommendationSummary_description :: Lens' RecommendationSummary (Maybe Text)

A description of the recommendation generated by CodeGuru Reviewer for the lines of code between the start line and the end line.

recommendationSummary_recommendationCategory :: Lens' RecommendationSummary (Maybe RecommendationCategory)

The type of a recommendation.

Repository

repository_codeCommit :: Lens' Repository (Maybe CodeCommitRepository)

Information about an Amazon Web Services CodeCommit repository.

repository_gitHubEnterpriseServer :: Lens' Repository (Maybe ThirdPartySourceRepository)

Information about a GitHub Enterprise Server repository.

repository_s3Bucket :: Lens' Repository (Maybe S3Repository)

Undocumented member.

repository_bitbucket :: Lens' Repository (Maybe ThirdPartySourceRepository)

Information about a Bitbucket repository.

RepositoryAnalysis

repositoryAnalysis_repositoryHead :: Lens' RepositoryAnalysis (Maybe RepositoryHeadSourceCodeType)

A SourceCodeType that specifies the tip of a branch in an associated repository.

repositoryAnalysis_sourceCodeType :: Lens' RepositoryAnalysis (Maybe SourceCodeType)

Undocumented member.

RepositoryAssociation

repositoryAssociation_associationArn :: Lens' RepositoryAssociation (Maybe Text)

The Amazon Resource Name (ARN) identifying the repository association.

repositoryAssociation_associationId :: Lens' RepositoryAssociation (Maybe Text)

The ID of the repository association.

repositoryAssociation_state :: Lens' RepositoryAssociation (Maybe RepositoryAssociationState)

The state of the repository association.

The valid repository association states are:

Associated: The repository association is complete.

Associating: CodeGuru Reviewer is:

Setting up pull request notifications. This is required for pull requests to trigger a CodeGuru Reviewer review.

If your repository ProviderType is GitHub, GitHub Enterprise Server, or Bitbucket, CodeGuru Reviewer creates webhooks in your repository to trigger CodeGuru Reviewer reviews. If you delete these webhooks, reviews of code in your repository cannot be triggered.

Setting up source code access. This is required for CodeGuru Reviewer to securely clone code in your repository.

Failed: The repository failed to associate or disassociate.

Disassociating: CodeGuru Reviewer is removing the repository's pull request notifications and source code access.

Disassociated: CodeGuru Reviewer successfully disassociated the repository. You can create a new association with this repository if you want to review source code in it later. You can control access to code reviews created in an associated repository with tags after it has been disassociated. For more information, see Using tags to control access to associated repositories in the Amazon CodeGuru Reviewer User Guide.

repositoryAssociation_s3RepositoryDetails :: Lens' RepositoryAssociation (Maybe S3RepositoryDetails)

Undocumented member.

repositoryAssociation_providerType :: Lens' RepositoryAssociation (Maybe ProviderType)

The provider type of the repository association.

repositoryAssociation_owner :: Lens' RepositoryAssociation (Maybe Text)

The owner of the repository. For an Amazon Web Services CodeCommit repository, this is the Amazon Web Services account ID of the account that owns the repository. For a GitHub, GitHub Enterprise Server, or Bitbucket repository, this is the username for the account that owns the repository. For an S3 repository, it can be the username or Amazon Web Services account ID.

repositoryAssociation_name :: Lens' RepositoryAssociation (Maybe Text)

The name of the repository.

repositoryAssociation_kmsKeyDetails :: Lens' RepositoryAssociation (Maybe KMSKeyDetails)

A KMSKeyDetails object that contains:

The encryption option for this repository association. It is either owned by Amazon Web Services Key Management Service (KMS) (AWS_OWNED_CMK) or customer managed (CUSTOMER_MANAGED_CMK).

The ID of the Amazon Web Services KMS key that is associated with this respository association.

repositoryAssociation_connectionArn :: Lens' RepositoryAssociation (Maybe Text)

The Amazon Resource Name (ARN) of an Amazon Web Services CodeStar Connections connection. Its format is arn:aws:codestar-connections:region-id:aws-account_id:connection/connection-id. For more information, see Connection in the Amazon Web Services CodeStar Connections API Reference.

repositoryAssociation_stateReason :: Lens' RepositoryAssociation (Maybe Text)

A description of why the repository association is in the current state.

repositoryAssociation_createdTimeStamp :: Lens' RepositoryAssociation (Maybe UTCTime)

The time, in milliseconds since the epoch, when the repository association was created.

repositoryAssociation_lastUpdatedTimeStamp :: Lens' RepositoryAssociation (Maybe UTCTime)

The time, in milliseconds since the epoch, when the repository association was last updated.

RepositoryAssociationSummary

repositoryAssociationSummary_associationArn :: Lens' RepositoryAssociationSummary (Maybe Text)

The Amazon Resource Name (ARN) of the RepositoryAssociation object. You can retrieve this ARN by calling ListRepositoryAssociations .

repositoryAssociationSummary_associationId :: Lens' RepositoryAssociationSummary (Maybe Text)

The repository association ID.

repositoryAssociationSummary_state :: Lens' RepositoryAssociationSummary (Maybe RepositoryAssociationState)

The state of the repository association.

The valid repository association states are:

Associated: The repository association is complete.

Associating: CodeGuru Reviewer is:

Setting up pull request notifications. This is required for pull requests to trigger a CodeGuru Reviewer review.

If your repository ProviderType is GitHub, GitHub Enterprise Server, or Bitbucket, CodeGuru Reviewer creates webhooks in your repository to trigger CodeGuru Reviewer reviews. If you delete these webhooks, reviews of code in your repository cannot be triggered.

Setting up source code access. This is required for CodeGuru Reviewer to securely clone code in your repository.

Failed: The repository failed to associate or disassociate.

Disassociating: CodeGuru Reviewer is removing the repository's pull request notifications and source code access.

Disassociated: CodeGuru Reviewer successfully disassociated the repository. You can create a new association with this repository if you want to review source code in it later. You can control access to code reviews created in an associated repository with tags after it has been disassociated. For more information, see Using tags to control access to associated repositories in the Amazon CodeGuru Reviewer User Guide.

repositoryAssociationSummary_providerType :: Lens' RepositoryAssociationSummary (Maybe ProviderType)

The provider type of the repository association.

repositoryAssociationSummary_owner :: Lens' RepositoryAssociationSummary (Maybe Text)

The owner of the repository. For an Amazon Web Services CodeCommit repository, this is the Amazon Web Services account ID of the account that owns the repository. For a GitHub, GitHub Enterprise Server, or Bitbucket repository, this is the username for the account that owns the repository. For an S3 repository, it can be the username or Amazon Web Services account ID.

repositoryAssociationSummary_name :: Lens' RepositoryAssociationSummary (Maybe Text)

The name of the repository association.

repositoryAssociationSummary_connectionArn :: Lens' RepositoryAssociationSummary (Maybe Text)

The Amazon Resource Name (ARN) of an Amazon Web Services CodeStar Connections connection. Its format is arn:aws:codestar-connections:region-id:aws-account_id:connection/connection-id. For more information, see Connection in the Amazon Web Services CodeStar Connections API Reference.

repositoryAssociationSummary_lastUpdatedTimeStamp :: Lens' RepositoryAssociationSummary (Maybe UTCTime)

The time, in milliseconds since the epoch, since the repository association was last updated.

RepositoryHeadSourceCodeType

repositoryHeadSourceCodeType_branchName :: Lens' RepositoryHeadSourceCodeType Text

The name of the branch in an associated repository. The RepositoryHeadSourceCodeType specifies the tip of this branch.

RequestMetadata

requestMetadata_requestId :: Lens' RequestMetadata (Maybe Text)

The ID of the request. This is required for a pull request code review.

requestMetadata_eventInfo :: Lens' RequestMetadata (Maybe EventInfo)

Information about the event associated with a code review.

requestMetadata_vendorName :: Lens' RequestMetadata (Maybe VendorName)

The name of the repository vendor used to upload code to an S3 bucket for a CI/CD code review. For example, if code and artifacts are uploaded to an S3 bucket for a CI/CD code review by GitHub scripts from a GitHub repository, then the repository association's ProviderType is S3Bucket and the CI/CD repository vendor name is GitHub. For more information, see the definition for ProviderType in RepositoryAssociation.

requestMetadata_requester :: Lens' RequestMetadata (Maybe Text)

An identifier, such as a name or account ID, that is associated with the requester. The Requester is used to capture the author/actor name of the event request.

RuleMetadata

ruleMetadata_longDescription :: Lens' RuleMetadata (Maybe Text)

A long description of the rule.

ruleMetadata_ruleTags :: Lens' RuleMetadata (Maybe (NonEmpty Text))

Tags that are associated with the rule.

ruleMetadata_ruleId :: Lens' RuleMetadata (Maybe Text)

The ID of the rule.

ruleMetadata_ruleName :: Lens' RuleMetadata (Maybe Text)

The name of the rule.

ruleMetadata_shortDescription :: Lens' RuleMetadata (Maybe Text)

A short description of the rule.

S3BucketRepository

s3BucketRepository_details :: Lens' S3BucketRepository (Maybe S3RepositoryDetails)

An S3RepositoryDetails object that specifies the name of an S3 bucket and a CodeArtifacts object. The CodeArtifacts object includes the S3 object keys for a source code .zip file and for a build artifacts .zip file.

s3BucketRepository_name :: Lens' S3BucketRepository Text

The name of the repository when the ProviderType is S3Bucket.

S3Repository

s3Repository_name :: Lens' S3Repository Text

The name of the repository in the S3 bucket.

s3Repository_bucketName :: Lens' S3Repository Text

The name of the S3 bucket used for associating a new S3 repository. It must begin with codeguru-reviewer-.

S3RepositoryDetails

s3RepositoryDetails_codeArtifacts :: Lens' S3RepositoryDetails (Maybe CodeArtifacts)

A CodeArtifacts object. The CodeArtifacts object includes the S3 object key for a source code .zip file and for a build artifacts .zip file that contains .jar or .class files.

s3RepositoryDetails_bucketName :: Lens' S3RepositoryDetails (Maybe Text)

The name of the S3 bucket used for associating a new S3 repository. It must begin with codeguru-reviewer-.

SourceCodeType

sourceCodeType_s3BucketRepository :: Lens' SourceCodeType (Maybe S3BucketRepository)

Information about an associated repository in an S3 bucket that includes its name and an S3RepositoryDetails object. The S3RepositoryDetails object includes the name of an S3 bucket, an S3 key for a source code .zip file, and an S3 key for a build artifacts .zip file. S3BucketRepository is required in SourceCodeType for S3BucketRepository based code reviews.

sourceCodeType_requestMetadata :: Lens' SourceCodeType (Maybe RequestMetadata)

Metadata that is associated with a code review. This applies to any type of code review supported by CodeGuru Reviewer. The RequestMetadaa field captures any event metadata. For example, it might capture metadata associated with an event trigger, such as a push or a pull request.

sourceCodeType_repositoryHead :: Lens' SourceCodeType (Maybe RepositoryHeadSourceCodeType)

Undocumented member.

sourceCodeType_commitDiff :: Lens' SourceCodeType (Maybe CommitDiffSourceCodeType)

A SourceCodeType that specifies a commit diff created by a pull request on an associated repository.

sourceCodeType_branchDiff :: Lens' SourceCodeType (Maybe BranchDiffSourceCodeType)

A type of SourceCodeType that specifies a source branch name and a destination branch name in an associated repository.

ThirdPartySourceRepository

thirdPartySourceRepository_name :: Lens' ThirdPartySourceRepository Text

The name of the third party source repository.

thirdPartySourceRepository_connectionArn :: Lens' ThirdPartySourceRepository Text

The Amazon Resource Name (ARN) of an Amazon Web Services CodeStar Connections connection. Its format is arn:aws:codestar-connections:region-id:aws-account_id:connection/connection-id. For more information, see Connection in the Amazon Web Services CodeStar Connections API Reference.

thirdPartySourceRepository_owner :: Lens' ThirdPartySourceRepository Text

The owner of the repository. For a GitHub, GitHub Enterprise, or Bitbucket repository, this is the username for the account that owns the repository. For an S3 repository, this can be the username or Amazon Web Services account ID.