Network.AWS.CodeGuruReviewer.Types.CodeReview

Description

Documentation

data CodeReview

Information about a code review. A code review belongs to the associated repository that contains the reviewed code.

See: newCodeReview smart constructor.

Constructors

CodeReview'

Fields

associationArn :: Maybe Text

The Amazon Resource Name (ARN) of the RepositoryAssociation that contains the reviewed source code. You can retrieve associated repository ARNs by calling ListRepositoryAssociations .

state :: Maybe JobState

The valid code review states are:

Completed: The code review is complete.

Pending: The code review started and has not completed or failed.

Failed: The code review failed.

Deleting: The code review is being deleted.

metrics :: Maybe Metrics

The statistics from the code review.

pullRequestId :: Maybe Text

The pull request ID for the code review.

providerType :: Maybe ProviderType

The type of repository that contains the reviewed code (for example, GitHub or Bitbucket).

owner :: Maybe Text

The owner of the repository. For an Amazon Web Services CodeCommit repository, this is the Amazon Web Services account ID of the account that owns the repository. For a GitHub, GitHub Enterprise Server, or Bitbucket repository, this is the username for the account that owns the repository. For an S3 repository, it can be the username or Amazon Web Services account ID.

name :: Maybe Text

The name of the code review.

codeReviewArn :: Maybe Text

The Amazon Resource Name (ARN) of the CodeReview object.

repositoryName :: Maybe Text

The name of the repository.

type' :: Maybe Type

The type of code review.

sourceCodeType :: Maybe SourceCodeType

The type of the source code for the code review.

stateReason :: Maybe Text

The reason for the state of the code review.

createdTimeStamp :: Maybe POSIX

The time, in milliseconds since the epoch, when the code review was created.

analysisTypes :: Maybe [AnalysisType]

They types of analysis performed during a repository analysis or a pull request review. You can specify either Security, CodeQuality, or both.

lastUpdatedTimeStamp :: Maybe POSIX

The time, in milliseconds since the epoch, when the code review was last updated.

newCodeReview :: CodeReview

Create a value of CodeReview with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:associationArn:CodeReview', codeReview_associationArn - The Amazon Resource Name (ARN) of the RepositoryAssociation that contains the reviewed source code. You can retrieve associated repository ARNs by calling ListRepositoryAssociations .

$sel:state:CodeReview', codeReview_state - The valid code review states are:

Completed: The code review is complete.

Pending: The code review started and has not completed or failed.

Failed: The code review failed.

Deleting: The code review is being deleted.

$sel:metrics:CodeReview', codeReview_metrics - The statistics from the code review.

$sel:pullRequestId:CodeReview', codeReview_pullRequestId - The pull request ID for the code review.

$sel:providerType:CodeReview', codeReview_providerType - The type of repository that contains the reviewed code (for example, GitHub or Bitbucket).

$sel:owner:CodeReview', codeReview_owner - The owner of the repository. For an Amazon Web Services CodeCommit repository, this is the Amazon Web Services account ID of the account that owns the repository. For a GitHub, GitHub Enterprise Server, or Bitbucket repository, this is the username for the account that owns the repository. For an S3 repository, it can be the username or Amazon Web Services account ID.

$sel:name:CodeReview', codeReview_name - The name of the code review.

$sel:codeReviewArn:CodeReview', codeReview_codeReviewArn - The Amazon Resource Name (ARN) of the CodeReview object.

$sel:repositoryName:CodeReview', codeReview_repositoryName - The name of the repository.

$sel:type':CodeReview', codeReview_type - The type of code review.

$sel:sourceCodeType:CodeReview', codeReview_sourceCodeType - The type of the source code for the code review.

$sel:stateReason:CodeReview', codeReview_stateReason - The reason for the state of the code review.

$sel:createdTimeStamp:CodeReview', codeReview_createdTimeStamp - The time, in milliseconds since the epoch, when the code review was created.

$sel:analysisTypes:CodeReview', codeReview_analysisTypes - They types of analysis performed during a repository analysis or a pull request review. You can specify either Security, CodeQuality, or both.

$sel:lastUpdatedTimeStamp:CodeReview', codeReview_lastUpdatedTimeStamp - The time, in milliseconds since the epoch, when the code review was last updated.

codeReview_associationArn :: Lens' CodeReview (Maybe Text)

The Amazon Resource Name (ARN) of the RepositoryAssociation that contains the reviewed source code. You can retrieve associated repository ARNs by calling ListRepositoryAssociations .

codeReview_state :: Lens' CodeReview (Maybe JobState)

The valid code review states are:

Completed: The code review is complete.

Pending: The code review started and has not completed or failed.

Failed: The code review failed.

Deleting: The code review is being deleted.

codeReview_metrics :: Lens' CodeReview (Maybe Metrics)

The statistics from the code review.

codeReview_pullRequestId :: Lens' CodeReview (Maybe Text)

The pull request ID for the code review.

codeReview_providerType :: Lens' CodeReview (Maybe ProviderType)

The type of repository that contains the reviewed code (for example, GitHub or Bitbucket).

codeReview_owner :: Lens' CodeReview (Maybe Text)

The owner of the repository. For an Amazon Web Services CodeCommit repository, this is the Amazon Web Services account ID of the account that owns the repository. For a GitHub, GitHub Enterprise Server, or Bitbucket repository, this is the username for the account that owns the repository. For an S3 repository, it can be the username or Amazon Web Services account ID.

codeReview_name :: Lens' CodeReview (Maybe Text)

The name of the code review.

codeReview_codeReviewArn :: Lens' CodeReview (Maybe Text)

The Amazon Resource Name (ARN) of the CodeReview object.

codeReview_repositoryName :: Lens' CodeReview (Maybe Text)

The name of the repository.

codeReview_type :: Lens' CodeReview (Maybe Type)

The type of code review.

codeReview_sourceCodeType :: Lens' CodeReview (Maybe SourceCodeType)

The type of the source code for the code review.

codeReview_stateReason :: Lens' CodeReview (Maybe Text)

The reason for the state of the code review.

codeReview_createdTimeStamp :: Lens' CodeReview (Maybe UTCTime)

The time, in milliseconds since the epoch, when the code review was created.

codeReview_analysisTypes :: Lens' CodeReview (Maybe [AnalysisType])

They types of analysis performed during a repository analysis or a pull request review. You can specify either Security, CodeQuality, or both.

codeReview_lastUpdatedTimeStamp :: Lens' CodeReview (Maybe UTCTime)

The time, in milliseconds since the epoch, when the code review was last updated.