Network.AWS.CodePipeline.CreateCustomActionType

Description

Creates a new custom action that can be used in all pipelines associated with the AWS account. Only used for custom actions.

Creating a Request

data CreateCustomActionType

Represents the input of a CreateCustomActionType operation.

See: newCreateCustomActionType smart constructor.

Constructors

CreateCustomActionType'

Fields

configurationProperties :: Maybe [ActionConfigurationProperty]

The configuration properties for the custom action.

You can refer to a name in the configuration properties of the custom action within the URL templates by following the format of {Config:name}, as long as the configuration property is both required and not secret. For more information, see Create a Custom Action for a Pipeline.

tags :: Maybe [Tag]

The tags for the custom action.

settings :: Maybe ActionTypeSettings

URLs that provide users information about this custom action.

category :: ActionCategory

The category of the custom action, such as a build action or a test action.

provider :: Text

The provider of the service used in the custom action, such as AWS CodeDeploy.

version :: Text

The version identifier of the custom action.

inputArtifactDetails :: ArtifactDetails

The details of the input artifact for the action, such as its commit ID.

outputArtifactDetails :: ArtifactDetails

The details of the output artifact of the action, such as its commit ID.

newCreateCustomActionType

Arguments

:: ActionCategory
$sel:category:CreateCustomActionType'

-> Text
$sel:provider:CreateCustomActionType'

-> Text
$sel:version:CreateCustomActionType'

-> ArtifactDetails
$sel:inputArtifactDetails:CreateCustomActionType'

-> ArtifactDetails
$sel:outputArtifactDetails:CreateCustomActionType'

-> CreateCustomActionType

Create a value of CreateCustomActionType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configurationProperties:CreateCustomActionType', createCustomActionType_configurationProperties - The configuration properties for the custom action.

You can refer to a name in the configuration properties of the custom action within the URL templates by following the format of {Config:name}, as long as the configuration property is both required and not secret. For more information, see Create a Custom Action for a Pipeline.

$sel:tags:CreateCustomActionType', createCustomActionType_tags - The tags for the custom action.

$sel:settings:CreateCustomActionType', createCustomActionType_settings - URLs that provide users information about this custom action.

$sel:category:CreateCustomActionType', createCustomActionType_category - The category of the custom action, such as a build action or a test action.

$sel:provider:CreateCustomActionType', createCustomActionType_provider - The provider of the service used in the custom action, such as AWS CodeDeploy.

$sel:version:CreateCustomActionType', createCustomActionType_version - The version identifier of the custom action.

$sel:inputArtifactDetails:CreateCustomActionType', createCustomActionType_inputArtifactDetails - The details of the input artifact for the action, such as its commit ID.

$sel:outputArtifactDetails:CreateCustomActionType', createCustomActionType_outputArtifactDetails - The details of the output artifact of the action, such as its commit ID.

Request Lenses

createCustomActionType_configurationProperties :: Lens' CreateCustomActionType (Maybe [ActionConfigurationProperty])

The configuration properties for the custom action.

You can refer to a name in the configuration properties of the custom action within the URL templates by following the format of {Config:name}, as long as the configuration property is both required and not secret. For more information, see Create a Custom Action for a Pipeline.

createCustomActionType_tags :: Lens' CreateCustomActionType (Maybe [Tag])

The tags for the custom action.

createCustomActionType_settings :: Lens' CreateCustomActionType (Maybe ActionTypeSettings)

URLs that provide users information about this custom action.

createCustomActionType_category :: Lens' CreateCustomActionType ActionCategory

The category of the custom action, such as a build action or a test action.

createCustomActionType_provider :: Lens' CreateCustomActionType Text

The provider of the service used in the custom action, such as AWS CodeDeploy.

createCustomActionType_version :: Lens' CreateCustomActionType Text

The version identifier of the custom action.

createCustomActionType_inputArtifactDetails :: Lens' CreateCustomActionType ArtifactDetails

The details of the input artifact for the action, such as its commit ID.

createCustomActionType_outputArtifactDetails :: Lens' CreateCustomActionType ArtifactDetails

The details of the output artifact of the action, such as its commit ID.

Destructuring the Response

data CreateCustomActionTypeResponse

Represents the output of a CreateCustomActionType operation.

See: newCreateCustomActionTypeResponse smart constructor.

Constructors

CreateCustomActionTypeResponse'

Fields

tags :: Maybe [Tag]

Specifies the tags applied to the custom action.

httpStatus :: Int

The response's http status code.

actionType :: ActionType

Returns information about the details of an action type.

newCreateCustomActionTypeResponse

Arguments

:: Int
$sel:httpStatus:CreateCustomActionTypeResponse'

-> ActionType
$sel:actionType:CreateCustomActionTypeResponse'

-> CreateCustomActionTypeResponse

Create a value of CreateCustomActionTypeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateCustomActionType', createCustomActionTypeResponse_tags - Specifies the tags applied to the custom action.

$sel:httpStatus:CreateCustomActionTypeResponse', createCustomActionTypeResponse_httpStatus - The response's http status code.

$sel:actionType:CreateCustomActionTypeResponse', createCustomActionTypeResponse_actionType - Returns information about the details of an action type.

Response Lenses

createCustomActionTypeResponse_tags :: Lens' CreateCustomActionTypeResponse (Maybe [Tag])

Specifies the tags applied to the custom action.

createCustomActionTypeResponse_httpStatus :: Lens' CreateCustomActionTypeResponse Int

The response's http status code.

createCustomActionTypeResponse_actionType :: Lens' CreateCustomActionTypeResponse ActionType

Returns information about the details of an action type.