Network.AWS.CodePipeline.GetActionType

Description

Returns information about an action type created for an external provider, where the action is to be used by customers of the external provider. The action can be created with any supported integration model.

Creating a Request

data GetActionType

See: newGetActionType smart constructor.

Constructors

GetActionType'

Fields

category :: ActionCategory

Defines what kind of action can be taken in the stage. The following are the valid values:

Source

Build

Test

Deploy

Approval

Invoke

owner :: Text

The creator of an action type that was created with any supported integration model. There are two valid values: AWS and ThirdParty.

provider :: Text

The provider of the action type being called. The provider name is specified when the action type is created.

version :: Text

A string that describes the action type version.

newGetActionType

Arguments

:: ActionCategory
$sel:category:GetActionType'

-> Text
$sel:owner:GetActionType'

-> Text
$sel:provider:GetActionType'

-> Text
$sel:version:GetActionType'

-> GetActionType

Create a value of GetActionType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:category:GetActionType', getActionType_category - Defines what kind of action can be taken in the stage. The following are the valid values:

Source

Build

Test

Deploy

Approval

Invoke

$sel:owner:GetActionType', getActionType_owner - The creator of an action type that was created with any supported integration model. There are two valid values: AWS and ThirdParty.

$sel:provider:GetActionType', getActionType_provider - The provider of the action type being called. The provider name is specified when the action type is created.

$sel:version:GetActionType', getActionType_version - A string that describes the action type version.

Request Lenses

getActionType_category :: Lens' GetActionType ActionCategory

Defines what kind of action can be taken in the stage. The following are the valid values:

Source

Build

Test

Deploy

Approval

Invoke

getActionType_owner :: Lens' GetActionType Text

The creator of an action type that was created with any supported integration model. There are two valid values: AWS and ThirdParty.

getActionType_provider :: Lens' GetActionType Text

The provider of the action type being called. The provider name is specified when the action type is created.

getActionType_version :: Lens' GetActionType Text

A string that describes the action type version.

Destructuring the Response

data GetActionTypeResponse

See: newGetActionTypeResponse smart constructor.

Constructors

GetActionTypeResponse'

Fields

actionType :: Maybe ActionTypeDeclaration

The action type information for the requested action type, such as the action type ID.

httpStatus :: Int

The response's http status code.

newGetActionTypeResponse

Arguments

:: Int
$sel:httpStatus:GetActionTypeResponse'

-> GetActionTypeResponse

Create a value of GetActionTypeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:actionType:GetActionTypeResponse', getActionTypeResponse_actionType - The action type information for the requested action type, such as the action type ID.

$sel:httpStatus:GetActionTypeResponse', getActionTypeResponse_httpStatus - The response's http status code.

Response Lenses

getActionTypeResponse_actionType :: Lens' GetActionTypeResponse (Maybe ActionTypeDeclaration)

The action type information for the requested action type, such as the action type ID.

getActionTypeResponse_httpStatus :: Lens' GetActionTypeResponse Int

The response's http status code.