Network.AWS.CodePipeline.GetPipeline

Description

Returns the metadata, structure, stages, and actions of a pipeline. Can be used to return the entire structure of a pipeline in JSON format, which can then be modified and used to update the pipeline structure with UpdatePipeline.

Creating a Request

data GetPipeline

Represents the input of a GetPipeline action.

See: newGetPipeline smart constructor.

Constructors

GetPipeline'

Fields

version :: Maybe Natural

The version number of the pipeline. If you do not specify a version, defaults to the current version.

name :: Text

The name of the pipeline for which you want to get information. Pipeline names must be unique under an AWS user account.

newGetPipeline

Arguments

:: Text
$sel:name:GetPipeline'

-> GetPipeline

Create a value of GetPipeline with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:version:GetPipeline', getPipeline_version - The version number of the pipeline. If you do not specify a version, defaults to the current version.

$sel:name:GetPipeline', getPipeline_name - The name of the pipeline for which you want to get information. Pipeline names must be unique under an AWS user account.

Request Lenses

getPipeline_version :: Lens' GetPipeline (Maybe Natural)

The version number of the pipeline. If you do not specify a version, defaults to the current version.

getPipeline_name :: Lens' GetPipeline Text

The name of the pipeline for which you want to get information. Pipeline names must be unique under an AWS user account.

Destructuring the Response

data GetPipelineResponse

Represents the output of a GetPipeline action.

See: newGetPipelineResponse smart constructor.

Constructors

GetPipelineResponse'

Fields

pipeline :: Maybe PipelineDeclaration

Represents the structure of actions and stages to be performed in the pipeline.

metadata :: Maybe PipelineMetadata

Represents the pipeline metadata information returned as part of the output of a GetPipeline action.

httpStatus :: Int

The response's http status code.

newGetPipelineResponse

Arguments

:: Int
$sel:httpStatus:GetPipelineResponse'

-> GetPipelineResponse

Create a value of GetPipelineResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pipeline:GetPipelineResponse', getPipelineResponse_pipeline - Represents the structure of actions and stages to be performed in the pipeline.

$sel:metadata:GetPipelineResponse', getPipelineResponse_metadata - Represents the pipeline metadata information returned as part of the output of a GetPipeline action.

$sel:httpStatus:GetPipelineResponse', getPipelineResponse_httpStatus - The response's http status code.

Response Lenses

getPipelineResponse_pipeline :: Lens' GetPipelineResponse (Maybe PipelineDeclaration)

Represents the structure of actions and stages to be performed in the pipeline.

getPipelineResponse_metadata :: Lens' GetPipelineResponse (Maybe PipelineMetadata)

Represents the pipeline metadata information returned as part of the output of a GetPipeline action.

getPipelineResponse_httpStatus :: Lens' GetPipelineResponse Int

The response's http status code.