Network.AWS.CodePipeline.ListActionExecutions

Description

Lists the action executions that have occurred in a pipeline.

This operation returns paginated results.

Creating a Request

data ListActionExecutions

See: newListActionExecutions smart constructor.

Constructors

ListActionExecutions'

Fields

nextToken :: Maybe Text

The token that was returned from the previous ListActionExecutions call, which can be used to return the next set of action executions in the list.

filter' :: Maybe ActionExecutionFilter

Input information used to filter action execution history.

maxResults :: Maybe Natural

The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned nextToken value. Action execution history is retained for up to 12 months, based on action execution start times. Default value is 100.

Detailed execution history is available for executions run on or after February 21, 2019.

pipelineName :: Text

The name of the pipeline for which you want to list action execution history.

newListActionExecutions

Arguments

:: Text
$sel:pipelineName:ListActionExecutions'

-> ListActionExecutions

Create a value of ListActionExecutions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListActionExecutions', listActionExecutions_nextToken - The token that was returned from the previous ListActionExecutions call, which can be used to return the next set of action executions in the list.

$sel:filter':ListActionExecutions', listActionExecutions_filter - Input information used to filter action execution history.

$sel:maxResults:ListActionExecutions', listActionExecutions_maxResults - The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned nextToken value. Action execution history is retained for up to 12 months, based on action execution start times. Default value is 100.

Detailed execution history is available for executions run on or after February 21, 2019.

$sel:pipelineName:ListActionExecutions', listActionExecutions_pipelineName - The name of the pipeline for which you want to list action execution history.

Request Lenses

listActionExecutions_nextToken :: Lens' ListActionExecutions (Maybe Text)

The token that was returned from the previous ListActionExecutions call, which can be used to return the next set of action executions in the list.

listActionExecutions_filter :: Lens' ListActionExecutions (Maybe ActionExecutionFilter)

Input information used to filter action execution history.

listActionExecutions_maxResults :: Lens' ListActionExecutions (Maybe Natural)

The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned nextToken value. Action execution history is retained for up to 12 months, based on action execution start times. Default value is 100.

Detailed execution history is available for executions run on or after February 21, 2019.

listActionExecutions_pipelineName :: Lens' ListActionExecutions Text

The name of the pipeline for which you want to list action execution history.

Destructuring the Response

data ListActionExecutionsResponse

See: newListActionExecutionsResponse smart constructor.

Constructors

ListActionExecutionsResponse'

Fields

actionExecutionDetails :: Maybe [ActionExecutionDetail]

The details for a list of recent executions, such as action execution ID.

nextToken :: Maybe Text

If the amount of returned information is significantly large, an identifier is also returned and can be used in a subsequent ListActionExecutions call to return the next set of action executions in the list.

httpStatus :: Int

The response's http status code.

newListActionExecutionsResponse

Arguments

:: Int
$sel:httpStatus:ListActionExecutionsResponse'

-> ListActionExecutionsResponse

Create a value of ListActionExecutionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:actionExecutionDetails:ListActionExecutionsResponse', listActionExecutionsResponse_actionExecutionDetails - The details for a list of recent executions, such as action execution ID.

$sel:nextToken:ListActionExecutions', listActionExecutionsResponse_nextToken - If the amount of returned information is significantly large, an identifier is also returned and can be used in a subsequent ListActionExecutions call to return the next set of action executions in the list.

$sel:httpStatus:ListActionExecutionsResponse', listActionExecutionsResponse_httpStatus - The response's http status code.

Response Lenses

listActionExecutionsResponse_actionExecutionDetails :: Lens' ListActionExecutionsResponse (Maybe [ActionExecutionDetail])

The details for a list of recent executions, such as action execution ID.

listActionExecutionsResponse_nextToken :: Lens' ListActionExecutionsResponse (Maybe Text)

If the amount of returned information is significantly large, an identifier is also returned and can be used in a subsequent ListActionExecutions call to return the next set of action executions in the list.

listActionExecutionsResponse_httpStatus :: Lens' ListActionExecutionsResponse Int

The response's http status code.