Network.AWS.CodePipeline.ListPipelineExecutions

Description

Gets a summary of the most recent executions for a pipeline.

This operation returns paginated results.

Creating a Request

data ListPipelineExecutions

Represents the input of a ListPipelineExecutions action.

See: newListPipelineExecutions smart constructor.

Constructors

ListPipelineExecutions'

Fields

nextToken :: Maybe Text

The token that was returned from the previous ListPipelineExecutions call, which can be used to return the next set of pipeline executions in the list.

maxResults :: Maybe Natural

The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned nextToken value. Pipeline history is limited to the most recent 12 months, based on pipeline execution start times. Default value is 100.

pipelineName :: Text

The name of the pipeline for which you want to get execution summary information.

newListPipelineExecutions

Arguments

:: Text
$sel:pipelineName:ListPipelineExecutions'

-> ListPipelineExecutions

Create a value of ListPipelineExecutions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListPipelineExecutions', listPipelineExecutions_nextToken - The token that was returned from the previous ListPipelineExecutions call, which can be used to return the next set of pipeline executions in the list.

$sel:maxResults:ListPipelineExecutions', listPipelineExecutions_maxResults - The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned nextToken value. Pipeline history is limited to the most recent 12 months, based on pipeline execution start times. Default value is 100.

$sel:pipelineName:ListPipelineExecutions', listPipelineExecutions_pipelineName - The name of the pipeline for which you want to get execution summary information.

Request Lenses

listPipelineExecutions_nextToken :: Lens' ListPipelineExecutions (Maybe Text)

The token that was returned from the previous ListPipelineExecutions call, which can be used to return the next set of pipeline executions in the list.

listPipelineExecutions_maxResults :: Lens' ListPipelineExecutions (Maybe Natural)

The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned nextToken value. Pipeline history is limited to the most recent 12 months, based on pipeline execution start times. Default value is 100.

listPipelineExecutions_pipelineName :: Lens' ListPipelineExecutions Text

The name of the pipeline for which you want to get execution summary information.

Destructuring the Response

data ListPipelineExecutionsResponse

Represents the output of a ListPipelineExecutions action.

See: newListPipelineExecutionsResponse smart constructor.

Constructors

ListPipelineExecutionsResponse'

Fields

nextToken :: Maybe Text

A token that can be used in the next ListPipelineExecutions call. To view all items in the list, continue to call this operation with each subsequent token until no more nextToken values are returned.

pipelineExecutionSummaries :: Maybe [PipelineExecutionSummary]

A list of executions in the history of a pipeline.

httpStatus :: Int

The response's http status code.

newListPipelineExecutionsResponse

Arguments

:: Int
$sel:httpStatus:ListPipelineExecutionsResponse'

-> ListPipelineExecutionsResponse

Create a value of ListPipelineExecutionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListPipelineExecutions', listPipelineExecutionsResponse_nextToken - A token that can be used in the next ListPipelineExecutions call. To view all items in the list, continue to call this operation with each subsequent token until no more nextToken values are returned.

$sel:pipelineExecutionSummaries:ListPipelineExecutionsResponse', listPipelineExecutionsResponse_pipelineExecutionSummaries - A list of executions in the history of a pipeline.

$sel:httpStatus:ListPipelineExecutionsResponse', listPipelineExecutionsResponse_httpStatus - The response's http status code.

Response Lenses

listPipelineExecutionsResponse_nextToken :: Lens' ListPipelineExecutionsResponse (Maybe Text)

A token that can be used in the next ListPipelineExecutions call. To view all items in the list, continue to call this operation with each subsequent token until no more nextToken values are returned.

listPipelineExecutionsResponse_pipelineExecutionSummaries :: Lens' ListPipelineExecutionsResponse (Maybe [PipelineExecutionSummary])

A list of executions in the history of a pipeline.

listPipelineExecutionsResponse_httpStatus :: Lens' ListPipelineExecutionsResponse Int

The response's http status code.