Network.AWS.CodePipeline.PutActionRevision

Description

Provides information to AWS CodePipeline about new revisions to a source.

Creating a Request

data PutActionRevision

Represents the input of a PutActionRevision action.

See: newPutActionRevision smart constructor.

Constructors

PutActionRevision'

Fields

pipelineName :: Text

The name of the pipeline that starts processing the revision to the source.

stageName :: Text

The name of the stage that contains the action that acts on the revision.

actionName :: Text

The name of the action that processes the revision.

actionRevision :: ActionRevision

Represents information about the version (or revision) of an action.

newPutActionRevision

Arguments

:: Text
$sel:pipelineName:PutActionRevision'

-> Text
$sel:stageName:PutActionRevision'

-> Text
$sel:actionName:PutActionRevision'

-> ActionRevision
$sel:actionRevision:PutActionRevision'

-> PutActionRevision

Create a value of PutActionRevision with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pipelineName:PutActionRevision', putActionRevision_pipelineName - The name of the pipeline that starts processing the revision to the source.

$sel:stageName:PutActionRevision', putActionRevision_stageName - The name of the stage that contains the action that acts on the revision.

$sel:actionName:PutActionRevision', putActionRevision_actionName - The name of the action that processes the revision.

$sel:actionRevision:PutActionRevision', putActionRevision_actionRevision - Represents information about the version (or revision) of an action.

Request Lenses

putActionRevision_pipelineName :: Lens' PutActionRevision Text

The name of the pipeline that starts processing the revision to the source.

putActionRevision_stageName :: Lens' PutActionRevision Text

The name of the stage that contains the action that acts on the revision.

putActionRevision_actionName :: Lens' PutActionRevision Text

The name of the action that processes the revision.

putActionRevision_actionRevision :: Lens' PutActionRevision ActionRevision

Represents information about the version (or revision) of an action.

Destructuring the Response

data PutActionRevisionResponse

Represents the output of a PutActionRevision action.

See: newPutActionRevisionResponse smart constructor.

Constructors

PutActionRevisionResponse'

Fields

newRevision' :: Maybe Bool

Indicates whether the artifact revision was previously used in an execution of the specified pipeline.

pipelineExecutionId :: Maybe Text

The ID of the current workflow state of the pipeline.

httpStatus :: Int

The response's http status code.

newPutActionRevisionResponse

Arguments

:: Int
$sel:httpStatus:PutActionRevisionResponse'

-> PutActionRevisionResponse

Create a value of PutActionRevisionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:newRevision':PutActionRevisionResponse', putActionRevisionResponse_newRevision - Indicates whether the artifact revision was previously used in an execution of the specified pipeline.

$sel:pipelineExecutionId:PutActionRevisionResponse', putActionRevisionResponse_pipelineExecutionId - The ID of the current workflow state of the pipeline.

$sel:httpStatus:PutActionRevisionResponse', putActionRevisionResponse_httpStatus - The response's http status code.

Response Lenses

putActionRevisionResponse_newRevision :: Lens' PutActionRevisionResponse (Maybe Bool)

Indicates whether the artifact revision was previously used in an execution of the specified pipeline.

putActionRevisionResponse_pipelineExecutionId :: Lens' PutActionRevisionResponse (Maybe Text)

The ID of the current workflow state of the pipeline.

putActionRevisionResponse_httpStatus :: Lens' PutActionRevisionResponse Int

The response's http status code.