Network.AWS.CodePipeline.PutApprovalResult

Description

Provides the response to a manual approval request to AWS CodePipeline. Valid responses include Approved and Rejected.

Creating a Request

data PutApprovalResult

Represents the input of a PutApprovalResult action.

See: newPutApprovalResult smart constructor.

Constructors

PutApprovalResult'

Fields

pipelineName :: Text

The name of the pipeline that contains the action.

stageName :: Text

The name of the stage that contains the action.

actionName :: Text

The name of the action for which approval is requested.

result :: ApprovalResult

Represents information about the result of the approval request.

token :: Text

The system-generated token used to identify a unique approval request. The token for each open approval request can be obtained using the GetPipelineState action. It is used to validate that the approval request corresponding to this token is still valid.

newPutApprovalResult

Arguments

:: Text
$sel:pipelineName:PutApprovalResult'

-> Text
$sel:stageName:PutApprovalResult'

-> Text
$sel:actionName:PutApprovalResult'

-> ApprovalResult
$sel:result:PutApprovalResult'

-> Text
$sel:token:PutApprovalResult'

-> PutApprovalResult

Create a value of PutApprovalResult with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pipelineName:PutApprovalResult', putApprovalResult_pipelineName - The name of the pipeline that contains the action.

$sel:stageName:PutApprovalResult', putApprovalResult_stageName - The name of the stage that contains the action.

$sel:actionName:PutApprovalResult', putApprovalResult_actionName - The name of the action for which approval is requested.

$sel:result:PutApprovalResult', putApprovalResult_result - Represents information about the result of the approval request.

$sel:token:PutApprovalResult', putApprovalResult_token - The system-generated token used to identify a unique approval request. The token for each open approval request can be obtained using the GetPipelineState action. It is used to validate that the approval request corresponding to this token is still valid.

Request Lenses

putApprovalResult_pipelineName :: Lens' PutApprovalResult Text

The name of the pipeline that contains the action.

putApprovalResult_stageName :: Lens' PutApprovalResult Text

The name of the stage that contains the action.

putApprovalResult_actionName :: Lens' PutApprovalResult Text

The name of the action for which approval is requested.

putApprovalResult_result :: Lens' PutApprovalResult ApprovalResult

Represents information about the result of the approval request.

putApprovalResult_token :: Lens' PutApprovalResult Text

The system-generated token used to identify a unique approval request. The token for each open approval request can be obtained using the GetPipelineState action. It is used to validate that the approval request corresponding to this token is still valid.

Destructuring the Response

data PutApprovalResultResponse

Represents the output of a PutApprovalResult action.

See: newPutApprovalResultResponse smart constructor.

Constructors

PutApprovalResultResponse'

Fields

approvedAt :: Maybe POSIX

The timestamp showing when the approval or rejection was submitted.

httpStatus :: Int

The response's http status code.

newPutApprovalResultResponse

Arguments

:: Int
$sel:httpStatus:PutApprovalResultResponse'

-> PutApprovalResultResponse

Create a value of PutApprovalResultResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:approvedAt:PutApprovalResultResponse', putApprovalResultResponse_approvedAt - The timestamp showing when the approval or rejection was submitted.

$sel:httpStatus:PutApprovalResultResponse', putApprovalResultResponse_httpStatus - The response's http status code.

Response Lenses

putApprovalResultResponse_approvedAt :: Lens' PutApprovalResultResponse (Maybe UTCTime)

The timestamp showing when the approval or rejection was submitted.

putApprovalResultResponse_httpStatus :: Lens' PutApprovalResultResponse Int

The response's http status code.