Network.AWS.CodePipeline.PutJobFailureResult

Description

Represents the failure of a job as returned to the pipeline by a job worker. Used for custom actions only.

Creating a Request

data PutJobFailureResult

Represents the input of a PutJobFailureResult action.

See: newPutJobFailureResult smart constructor.

Constructors

PutJobFailureResult'

Fields

jobId :: Text

The unique system-generated ID of the job that failed. This is the same ID returned from PollForJobs.

failureDetails :: FailureDetails

The details about the failure of a job.

newPutJobFailureResult

Arguments

:: Text
$sel:jobId:PutJobFailureResult'

-> FailureDetails
$sel:failureDetails:PutJobFailureResult'

-> PutJobFailureResult

Create a value of PutJobFailureResult with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobId:PutJobFailureResult', putJobFailureResult_jobId - The unique system-generated ID of the job that failed. This is the same ID returned from PollForJobs.

$sel:failureDetails:PutJobFailureResult', putJobFailureResult_failureDetails - The details about the failure of a job.

Request Lenses

putJobFailureResult_jobId :: Lens' PutJobFailureResult Text

The unique system-generated ID of the job that failed. This is the same ID returned from PollForJobs.

putJobFailureResult_failureDetails :: Lens' PutJobFailureResult FailureDetails

The details about the failure of a job.

Destructuring the Response

data PutJobFailureResultResponse

See: newPutJobFailureResultResponse smart constructor.

Constructors

PutJobFailureResultResponse'

newPutJobFailureResultResponse :: PutJobFailureResultResponse

Create a value of PutJobFailureResultResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.