Network.AWS.CodePipeline.PutJobSuccessResult

Description

Represents the success of a job as returned to the pipeline by a job worker. Used for custom actions only.

Creating a Request

data PutJobSuccessResult

Represents the input of a PutJobSuccessResult action.

See: newPutJobSuccessResult smart constructor.

Constructors

PutJobSuccessResult'

Fields

executionDetails :: Maybe ExecutionDetails

The execution details of the successful job, such as the actions taken by the job worker.

currentRevision :: Maybe CurrentRevision

The ID of the current revision of the artifact successfully worked on by the job.

continuationToken :: Maybe Text

A token generated by a job worker, such as an AWS CodeDeploy deployment ID, that a successful job provides to identify a custom action in progress. Future jobs use this token to identify the running instance of the action. It can be reused to return more information about the progress of the custom action. When the action is complete, no continuation token should be supplied.

outputVariables :: Maybe (HashMap Text Text)

Key-value pairs produced as output by a job worker that can be made available to a downstream action configuration. outputVariables can be included only when there is no continuation token on the request.

jobId :: Text

The unique system-generated ID of the job that succeeded. This is the same ID returned from PollForJobs.

newPutJobSuccessResult

Arguments

:: Text
$sel:jobId:PutJobSuccessResult'

-> PutJobSuccessResult

Create a value of PutJobSuccessResult with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:executionDetails:PutJobSuccessResult', putJobSuccessResult_executionDetails - The execution details of the successful job, such as the actions taken by the job worker.

$sel:currentRevision:PutJobSuccessResult', putJobSuccessResult_currentRevision - The ID of the current revision of the artifact successfully worked on by the job.

$sel:continuationToken:PutJobSuccessResult', putJobSuccessResult_continuationToken - A token generated by a job worker, such as an AWS CodeDeploy deployment ID, that a successful job provides to identify a custom action in progress. Future jobs use this token to identify the running instance of the action. It can be reused to return more information about the progress of the custom action. When the action is complete, no continuation token should be supplied.

$sel:outputVariables:PutJobSuccessResult', putJobSuccessResult_outputVariables - Key-value pairs produced as output by a job worker that can be made available to a downstream action configuration. outputVariables can be included only when there is no continuation token on the request.

$sel:jobId:PutJobSuccessResult', putJobSuccessResult_jobId - The unique system-generated ID of the job that succeeded. This is the same ID returned from PollForJobs.

Request Lenses

putJobSuccessResult_executionDetails :: Lens' PutJobSuccessResult (Maybe ExecutionDetails)

The execution details of the successful job, such as the actions taken by the job worker.

putJobSuccessResult_currentRevision :: Lens' PutJobSuccessResult (Maybe CurrentRevision)

The ID of the current revision of the artifact successfully worked on by the job.

putJobSuccessResult_continuationToken :: Lens' PutJobSuccessResult (Maybe Text)

A token generated by a job worker, such as an AWS CodeDeploy deployment ID, that a successful job provides to identify a custom action in progress. Future jobs use this token to identify the running instance of the action. It can be reused to return more information about the progress of the custom action. When the action is complete, no continuation token should be supplied.

putJobSuccessResult_outputVariables :: Lens' PutJobSuccessResult (Maybe (HashMap Text Text))

Key-value pairs produced as output by a job worker that can be made available to a downstream action configuration. outputVariables can be included only when there is no continuation token on the request.

putJobSuccessResult_jobId :: Lens' PutJobSuccessResult Text

The unique system-generated ID of the job that succeeded. This is the same ID returned from PollForJobs.

Destructuring the Response

data PutJobSuccessResultResponse

See: newPutJobSuccessResultResponse smart constructor.

Constructors

PutJobSuccessResultResponse'

newPutJobSuccessResultResponse :: PutJobSuccessResultResponse

Create a value of PutJobSuccessResultResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.