Network.AWS.CodePipeline.PutThirdPartyJobSuccessResult

Description

Represents the success of a third party job as returned to the pipeline by a job worker. Used for partner actions only.

Creating a Request

data PutThirdPartyJobSuccessResult

Represents the input of a PutThirdPartyJobSuccessResult action.

See: newPutThirdPartyJobSuccessResult smart constructor.

Constructors

PutThirdPartyJobSuccessResult'

Fields

executionDetails :: Maybe ExecutionDetails

The details of the actions taken and results produced on an artifact as it passes through stages in the pipeline.

currentRevision :: Maybe CurrentRevision

Represents information about a current revision.

continuationToken :: Maybe Text

A token generated by a job worker, such as an AWS CodeDeploy deployment ID, that a successful job provides to identify a partner action in progress. Future jobs use this token to identify the running instance of the action. It can be reused to return more information about the progress of the partner action. When the action is complete, no continuation token should be supplied.

jobId :: Text

The ID of the job that successfully completed. This is the same ID returned from PollForThirdPartyJobs.

clientToken :: Text

The clientToken portion of the clientId and clientToken pair used to verify that the calling entity is allowed access to the job and its details.

newPutThirdPartyJobSuccessResult

Arguments

:: Text
$sel:jobId:PutThirdPartyJobSuccessResult'

-> Text
$sel:clientToken:PutThirdPartyJobSuccessResult'

-> PutThirdPartyJobSuccessResult

Create a value of PutThirdPartyJobSuccessResult with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:executionDetails:PutThirdPartyJobSuccessResult', putThirdPartyJobSuccessResult_executionDetails - The details of the actions taken and results produced on an artifact as it passes through stages in the pipeline.

$sel:currentRevision:PutThirdPartyJobSuccessResult', putThirdPartyJobSuccessResult_currentRevision - Represents information about a current revision.

$sel:continuationToken:PutThirdPartyJobSuccessResult', putThirdPartyJobSuccessResult_continuationToken - A token generated by a job worker, such as an AWS CodeDeploy deployment ID, that a successful job provides to identify a partner action in progress. Future jobs use this token to identify the running instance of the action. It can be reused to return more information about the progress of the partner action. When the action is complete, no continuation token should be supplied.

$sel:jobId:PutThirdPartyJobSuccessResult', putThirdPartyJobSuccessResult_jobId - The ID of the job that successfully completed. This is the same ID returned from PollForThirdPartyJobs.

$sel:clientToken:PutThirdPartyJobSuccessResult', putThirdPartyJobSuccessResult_clientToken - The clientToken portion of the clientId and clientToken pair used to verify that the calling entity is allowed access to the job and its details.

Request Lenses

putThirdPartyJobSuccessResult_executionDetails :: Lens' PutThirdPartyJobSuccessResult (Maybe ExecutionDetails)

The details of the actions taken and results produced on an artifact as it passes through stages in the pipeline.

putThirdPartyJobSuccessResult_currentRevision :: Lens' PutThirdPartyJobSuccessResult (Maybe CurrentRevision)

Represents information about a current revision.

putThirdPartyJobSuccessResult_continuationToken :: Lens' PutThirdPartyJobSuccessResult (Maybe Text)

A token generated by a job worker, such as an AWS CodeDeploy deployment ID, that a successful job provides to identify a partner action in progress. Future jobs use this token to identify the running instance of the action. It can be reused to return more information about the progress of the partner action. When the action is complete, no continuation token should be supplied.

putThirdPartyJobSuccessResult_jobId :: Lens' PutThirdPartyJobSuccessResult Text

The ID of the job that successfully completed. This is the same ID returned from PollForThirdPartyJobs.

putThirdPartyJobSuccessResult_clientToken :: Lens' PutThirdPartyJobSuccessResult Text

The clientToken portion of the clientId and clientToken pair used to verify that the calling entity is allowed access to the job and its details.

Destructuring the Response

data PutThirdPartyJobSuccessResultResponse

See: newPutThirdPartyJobSuccessResultResponse smart constructor.

Constructors

PutThirdPartyJobSuccessResultResponse'

newPutThirdPartyJobSuccessResultResponse :: PutThirdPartyJobSuccessResultResponse

Create a value of PutThirdPartyJobSuccessResultResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.