Network.AWS.CodePipeline.RetryStageExecution

Description

Resumes the pipeline execution by retrying the last failed actions in a stage. You can retry a stage immediately if any of the actions in the stage fail. When you retry, all actions that are still in progress continue working, and failed actions are triggered again.

Creating a Request

data RetryStageExecution

Represents the input of a RetryStageExecution action.

See: newRetryStageExecution smart constructor.

Constructors

RetryStageExecution'

Fields

pipelineName :: Text

The name of the pipeline that contains the failed stage.

stageName :: Text

The name of the failed stage to be retried.

pipelineExecutionId :: Text

The ID of the pipeline execution in the failed stage to be retried. Use the GetPipelineState action to retrieve the current pipelineExecutionId of the failed stage

retryMode :: StageRetryMode

The scope of the retry attempt. Currently, the only supported value is FAILED_ACTIONS.

newRetryStageExecution

Arguments

:: Text
$sel:pipelineName:RetryStageExecution'

-> Text
$sel:stageName:RetryStageExecution'

-> Text
$sel:pipelineExecutionId:RetryStageExecution'

-> StageRetryMode
$sel:retryMode:RetryStageExecution'

-> RetryStageExecution

Create a value of RetryStageExecution with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pipelineName:RetryStageExecution', retryStageExecution_pipelineName - The name of the pipeline that contains the failed stage.

$sel:stageName:RetryStageExecution', retryStageExecution_stageName - The name of the failed stage to be retried.

$sel:pipelineExecutionId:RetryStageExecution', retryStageExecution_pipelineExecutionId - The ID of the pipeline execution in the failed stage to be retried. Use the GetPipelineState action to retrieve the current pipelineExecutionId of the failed stage

$sel:retryMode:RetryStageExecution', retryStageExecution_retryMode - The scope of the retry attempt. Currently, the only supported value is FAILED_ACTIONS.

Request Lenses

retryStageExecution_pipelineName :: Lens' RetryStageExecution Text

The name of the pipeline that contains the failed stage.

retryStageExecution_stageName :: Lens' RetryStageExecution Text

The name of the failed stage to be retried.

retryStageExecution_pipelineExecutionId :: Lens' RetryStageExecution Text

The ID of the pipeline execution in the failed stage to be retried. Use the GetPipelineState action to retrieve the current pipelineExecutionId of the failed stage

retryStageExecution_retryMode :: Lens' RetryStageExecution StageRetryMode

The scope of the retry attempt. Currently, the only supported value is FAILED_ACTIONS.

Destructuring the Response

data RetryStageExecutionResponse

Represents the output of a RetryStageExecution action.

See: newRetryStageExecutionResponse smart constructor.

Constructors

RetryStageExecutionResponse'

Fields

pipelineExecutionId :: Maybe Text

The ID of the current workflow execution in the failed stage.

httpStatus :: Int

The response's http status code.

newRetryStageExecutionResponse

Arguments

:: Int
$sel:httpStatus:RetryStageExecutionResponse'

-> RetryStageExecutionResponse

Create a value of RetryStageExecutionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pipelineExecutionId:RetryStageExecution', retryStageExecutionResponse_pipelineExecutionId - The ID of the current workflow execution in the failed stage.

$sel:httpStatus:RetryStageExecutionResponse', retryStageExecutionResponse_httpStatus - The response's http status code.

Response Lenses

retryStageExecutionResponse_pipelineExecutionId :: Lens' RetryStageExecutionResponse (Maybe Text)

The ID of the current workflow execution in the failed stage.

retryStageExecutionResponse_httpStatus :: Lens' RetryStageExecutionResponse Int

The response's http status code.