Network.AWS.CodePipeline.Types.AWSSessionCredentials

Description

Documentation

data AWSSessionCredentials

Represents an AWS session credentials object. These credentials are temporary credentials that are issued by AWS Secure Token Service (STS). They can be used to access input and output artifacts in the S3 bucket used to store artifact for the pipeline in AWS CodePipeline.

See: newAWSSessionCredentials smart constructor.

Constructors

AWSSessionCredentials'

Fields

accessKeyId :: Sensitive Text

The access key for the session.

secretAccessKey :: Sensitive Text

The secret access key for the session.

sessionToken :: Sensitive Text

The token for the session.

newAWSSessionCredentials

Arguments

:: Text
$sel:accessKeyId:AWSSessionCredentials'

-> Text
$sel:secretAccessKey:AWSSessionCredentials'

-> Text
$sel:sessionToken:AWSSessionCredentials'

-> AWSSessionCredentials

Create a value of AWSSessionCredentials with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accessKeyId:AWSSessionCredentials', aWSSessionCredentials_accessKeyId - The access key for the session.

$sel:secretAccessKey:AWSSessionCredentials', aWSSessionCredentials_secretAccessKey - The secret access key for the session.

$sel:sessionToken:AWSSessionCredentials', aWSSessionCredentials_sessionToken - The token for the session.

aWSSessionCredentials_accessKeyId :: Lens' AWSSessionCredentials Text

The access key for the session.

aWSSessionCredentials_secretAccessKey :: Lens' AWSSessionCredentials Text

The secret access key for the session.

aWSSessionCredentials_sessionToken :: Lens' AWSSessionCredentials Text

The token for the session.