Network.AWS.CodePipeline.Types.ActionConfigurationProperty

Description

Documentation

data ActionConfigurationProperty

Represents information about an action configuration property.

See: newActionConfigurationProperty smart constructor.

Constructors

ActionConfigurationProperty'

Fields

description :: Maybe Text

The description of the action configuration property that is displayed to users.

type' :: Maybe ActionConfigurationPropertyType

The type of the configuration property.

queryable :: Maybe Bool

Indicates that the property is used with PollForJobs. When creating a custom action, an action can have up to one queryable property. If it has one, that property must be both required and not secret.

If you create a pipeline with a custom action type, and that custom action contains a queryable property, the value for that configuration property is subject to other restrictions. The value must be less than or equal to twenty (20) characters. The value can contain only alphanumeric characters, underscores, and hyphens.

name :: Text

The name of the action configuration property.

required :: Bool

Whether the configuration property is a required value.

key :: Bool

Whether the configuration property is a key.

secret :: Bool

Whether the configuration property is secret. Secrets are hidden from all calls except for GetJobDetails, GetThirdPartyJobDetails, PollForJobs, and PollForThirdPartyJobs.

When updating a pipeline, passing * * * * * without changing any other values of the action preserves the previous value of the secret.

newActionConfigurationProperty

Arguments

:: Text
$sel:name:ActionConfigurationProperty'

-> Bool
$sel:required:ActionConfigurationProperty'

-> Bool
$sel:key:ActionConfigurationProperty'

-> Bool
$sel:secret:ActionConfigurationProperty'

-> ActionConfigurationProperty

Create a value of ActionConfigurationProperty with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:ActionConfigurationProperty', actionConfigurationProperty_description - The description of the action configuration property that is displayed to users.

$sel:type':ActionConfigurationProperty', actionConfigurationProperty_type - The type of the configuration property.

$sel:queryable:ActionConfigurationProperty', actionConfigurationProperty_queryable - Indicates that the property is used with PollForJobs. When creating a custom action, an action can have up to one queryable property. If it has one, that property must be both required and not secret.

If you create a pipeline with a custom action type, and that custom action contains a queryable property, the value for that configuration property is subject to other restrictions. The value must be less than or equal to twenty (20) characters. The value can contain only alphanumeric characters, underscores, and hyphens.

$sel:name:ActionConfigurationProperty', actionConfigurationProperty_name - The name of the action configuration property.

$sel:required:ActionConfigurationProperty', actionConfigurationProperty_required - Whether the configuration property is a required value.

$sel:key:ActionConfigurationProperty', actionConfigurationProperty_key - Whether the configuration property is a key.

$sel:secret:ActionConfigurationProperty', actionConfigurationProperty_secret - Whether the configuration property is secret. Secrets are hidden from all calls except for GetJobDetails, GetThirdPartyJobDetails, PollForJobs, and PollForThirdPartyJobs.

When updating a pipeline, passing * * * * * without changing any other values of the action preserves the previous value of the secret.

actionConfigurationProperty_description :: Lens' ActionConfigurationProperty (Maybe Text)

The description of the action configuration property that is displayed to users.

actionConfigurationProperty_type :: Lens' ActionConfigurationProperty (Maybe ActionConfigurationPropertyType)

The type of the configuration property.

actionConfigurationProperty_queryable :: Lens' ActionConfigurationProperty (Maybe Bool)

Indicates that the property is used with PollForJobs. When creating a custom action, an action can have up to one queryable property. If it has one, that property must be both required and not secret.

If you create a pipeline with a custom action type, and that custom action contains a queryable property, the value for that configuration property is subject to other restrictions. The value must be less than or equal to twenty (20) characters. The value can contain only alphanumeric characters, underscores, and hyphens.

actionConfigurationProperty_name :: Lens' ActionConfigurationProperty Text

The name of the action configuration property.

actionConfigurationProperty_required :: Lens' ActionConfigurationProperty Bool

Whether the configuration property is a required value.

actionConfigurationProperty_key :: Lens' ActionConfigurationProperty Bool

Whether the configuration property is a key.

actionConfigurationProperty_secret :: Lens' ActionConfigurationProperty Bool

Whether the configuration property is secret. Secrets are hidden from all calls except for GetJobDetails, GetThirdPartyJobDetails, PollForJobs, and PollForThirdPartyJobs.

When updating a pipeline, passing * * * * * without changing any other values of the action preserves the previous value of the secret.