Network.AWS.CodePipeline.Types.ActionDeclaration

Description

Documentation

data ActionDeclaration

Represents information about an action declaration.

See: newActionDeclaration smart constructor.

Constructors

ActionDeclaration'

Fields

roleArn :: Maybe Text

The ARN of the IAM service role that performs the declared action. This is assumed through the roleArn for the pipeline.

configuration :: Maybe (HashMap Text Text)

The action's configuration. These are key-value pairs that specify input values for an action. For more information, see Action Structure Requirements in CodePipeline. For the list of configuration properties for the AWS CloudFormation action type in CodePipeline, see Configuration Properties Reference in the AWS CloudFormation User Guide. For template snippets with examples, see Using Parameter Override Functions with CodePipeline Pipelines in the AWS CloudFormation User Guide.

The values can be represented in either JSON or YAML format. For example, the JSON configuration item format is as follows:

JSON:

"Configuration" : { Key : Value },

runOrder :: Maybe Natural

The order in which actions are run.

inputArtifacts :: Maybe [InputArtifact]

The name or ID of the artifact consumed by the action, such as a test or build artifact.

namespace :: Maybe Text

The variable namespace associated with the action. All variables produced as output by this action fall under this namespace.

region :: Maybe Text

The action declaration's AWS Region, such as us-east-1.

outputArtifacts :: Maybe [OutputArtifact]

The name or ID of the result of the action declaration, such as a test or build artifact.

name :: Text

The action declaration's name.

actionTypeId :: ActionTypeId

Specifies the action type and the provider of the action.

newActionDeclaration

Arguments

:: Text
$sel:name:ActionDeclaration'

-> ActionTypeId
$sel:actionTypeId:ActionDeclaration'

-> ActionDeclaration

Create a value of ActionDeclaration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:roleArn:ActionDeclaration', actionDeclaration_roleArn - The ARN of the IAM service role that performs the declared action. This is assumed through the roleArn for the pipeline.

$sel:configuration:ActionDeclaration', actionDeclaration_configuration - The action's configuration. These are key-value pairs that specify input values for an action. For more information, see Action Structure Requirements in CodePipeline. For the list of configuration properties for the AWS CloudFormation action type in CodePipeline, see Configuration Properties Reference in the AWS CloudFormation User Guide. For template snippets with examples, see Using Parameter Override Functions with CodePipeline Pipelines in the AWS CloudFormation User Guide.

The values can be represented in either JSON or YAML format. For example, the JSON configuration item format is as follows:

JSON:

"Configuration" : { Key : Value },

$sel:runOrder:ActionDeclaration', actionDeclaration_runOrder - The order in which actions are run.

$sel:inputArtifacts:ActionDeclaration', actionDeclaration_inputArtifacts - The name or ID of the artifact consumed by the action, such as a test or build artifact.

$sel:namespace:ActionDeclaration', actionDeclaration_namespace - The variable namespace associated with the action. All variables produced as output by this action fall under this namespace.

$sel:region:ActionDeclaration', actionDeclaration_region - The action declaration's AWS Region, such as us-east-1.

$sel:outputArtifacts:ActionDeclaration', actionDeclaration_outputArtifacts - The name or ID of the result of the action declaration, such as a test or build artifact.

$sel:name:ActionDeclaration', actionDeclaration_name - The action declaration's name.

$sel:actionTypeId:ActionDeclaration', actionDeclaration_actionTypeId - Specifies the action type and the provider of the action.

actionDeclaration_roleArn :: Lens' ActionDeclaration (Maybe Text)

The ARN of the IAM service role that performs the declared action. This is assumed through the roleArn for the pipeline.

actionDeclaration_configuration :: Lens' ActionDeclaration (Maybe (HashMap Text Text))

The action's configuration. These are key-value pairs that specify input values for an action. For more information, see Action Structure Requirements in CodePipeline. For the list of configuration properties for the AWS CloudFormation action type in CodePipeline, see Configuration Properties Reference in the AWS CloudFormation User Guide. For template snippets with examples, see Using Parameter Override Functions with CodePipeline Pipelines in the AWS CloudFormation User Guide.

The values can be represented in either JSON or YAML format. For example, the JSON configuration item format is as follows:

JSON:

"Configuration" : { Key : Value },

actionDeclaration_runOrder :: Lens' ActionDeclaration (Maybe Natural)

The order in which actions are run.

actionDeclaration_inputArtifacts :: Lens' ActionDeclaration (Maybe [InputArtifact])

The name or ID of the artifact consumed by the action, such as a test or build artifact.

actionDeclaration_namespace :: Lens' ActionDeclaration (Maybe Text)

The variable namespace associated with the action. All variables produced as output by this action fall under this namespace.

actionDeclaration_region :: Lens' ActionDeclaration (Maybe Text)

The action declaration's AWS Region, such as us-east-1.

actionDeclaration_outputArtifacts :: Lens' ActionDeclaration (Maybe [OutputArtifact])

The name or ID of the result of the action declaration, such as a test or build artifact.

actionDeclaration_name :: Lens' ActionDeclaration Text

The action declaration's name.

actionDeclaration_actionTypeId :: Lens' ActionDeclaration ActionTypeId

Specifies the action type and the provider of the action.