Network.AWS.CodePipeline.Types.ActionExecution

Description

Documentation

data ActionExecution

Represents information about the run of an action.

See: newActionExecution smart constructor.

Constructors

ActionExecution'

Fields

status :: Maybe ActionExecutionStatus

The status of the action, or for a completed action, the last status of the action.

actionExecutionId :: Maybe Text

ID of the workflow action execution in the current stage. Use the GetPipelineState action to retrieve the current action execution details of the current stage.

For older executions, this field might be empty. The action execution ID is available for executions run on or after March 2020.

lastStatusChange :: Maybe POSIX

The last status change of the action.

percentComplete :: Maybe Natural

A percentage of completeness of the action as it runs.

externalExecutionId :: Maybe Text

The external ID of the run of the action.

externalExecutionUrl :: Maybe Text

The URL of a resource external to AWS that is used when running the action (for example, an external repository URL).

lastUpdatedBy :: Maybe Text

The ARN of the user who last changed the pipeline.

summary :: Maybe Text

A summary of the run of the action.

token :: Maybe Text

The system-generated token used to identify a unique approval request. The token for each open approval request can be obtained using the GetPipelineState command. It is used to validate that the approval request corresponding to this token is still valid.

errorDetails :: Maybe ErrorDetails

The details of an error returned by a URL external to AWS.

newActionExecution :: ActionExecution

Create a value of ActionExecution with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ActionExecution', actionExecution_status - The status of the action, or for a completed action, the last status of the action.

$sel:actionExecutionId:ActionExecution', actionExecution_actionExecutionId - ID of the workflow action execution in the current stage. Use the GetPipelineState action to retrieve the current action execution details of the current stage.

For older executions, this field might be empty. The action execution ID is available for executions run on or after March 2020.

$sel:lastStatusChange:ActionExecution', actionExecution_lastStatusChange - The last status change of the action.

$sel:percentComplete:ActionExecution', actionExecution_percentComplete - A percentage of completeness of the action as it runs.

$sel:externalExecutionId:ActionExecution', actionExecution_externalExecutionId - The external ID of the run of the action.

$sel:externalExecutionUrl:ActionExecution', actionExecution_externalExecutionUrl - The URL of a resource external to AWS that is used when running the action (for example, an external repository URL).

$sel:lastUpdatedBy:ActionExecution', actionExecution_lastUpdatedBy - The ARN of the user who last changed the pipeline.

$sel:summary:ActionExecution', actionExecution_summary - A summary of the run of the action.

$sel:token:ActionExecution', actionExecution_token - The system-generated token used to identify a unique approval request. The token for each open approval request can be obtained using the GetPipelineState command. It is used to validate that the approval request corresponding to this token is still valid.

$sel:errorDetails:ActionExecution', actionExecution_errorDetails - The details of an error returned by a URL external to AWS.

actionExecution_status :: Lens' ActionExecution (Maybe ActionExecutionStatus)

The status of the action, or for a completed action, the last status of the action.

actionExecution_actionExecutionId :: Lens' ActionExecution (Maybe Text)

ID of the workflow action execution in the current stage. Use the GetPipelineState action to retrieve the current action execution details of the current stage.

For older executions, this field might be empty. The action execution ID is available for executions run on or after March 2020.

actionExecution_lastStatusChange :: Lens' ActionExecution (Maybe UTCTime)

The last status change of the action.

actionExecution_percentComplete :: Lens' ActionExecution (Maybe Natural)

A percentage of completeness of the action as it runs.

actionExecution_externalExecutionId :: Lens' ActionExecution (Maybe Text)

The external ID of the run of the action.

actionExecution_externalExecutionUrl :: Lens' ActionExecution (Maybe Text)

The URL of a resource external to AWS that is used when running the action (for example, an external repository URL).

actionExecution_lastUpdatedBy :: Lens' ActionExecution (Maybe Text)

The ARN of the user who last changed the pipeline.

actionExecution_summary :: Lens' ActionExecution (Maybe Text)

A summary of the run of the action.

actionExecution_token :: Lens' ActionExecution (Maybe Text)

The system-generated token used to identify a unique approval request. The token for each open approval request can be obtained using the GetPipelineState command. It is used to validate that the approval request corresponding to this token is still valid.

actionExecution_errorDetails :: Lens' ActionExecution (Maybe ErrorDetails)

The details of an error returned by a URL external to AWS.