Network.AWS.CodePipeline.Types.ActionExecutionInput

Description

Documentation

data ActionExecutionInput

Input information used for an action execution.

See: newActionExecutionInput smart constructor.

Constructors

ActionExecutionInput'

Fields

namespace :: Maybe Text

The variable namespace associated with the action. All variables produced as output by this action fall under this namespace.

resolvedConfiguration :: Maybe (HashMap Text Text)

Configuration data for an action execution with all variable references replaced with their real values for the execution.

region :: Maybe Text

The AWS Region for the action, such as us-east-1.

configuration :: Maybe (HashMap Text Text)

Configuration data for an action execution.

actionTypeId :: Maybe ActionTypeId

inputArtifacts :: Maybe [ArtifactDetail]

Details of input artifacts of the action that correspond to the action execution.

roleArn :: Maybe Text

The ARN of the IAM service role that performs the declared action. This is assumed through the roleArn for the pipeline.

newActionExecutionInput :: ActionExecutionInput

Create a value of ActionExecutionInput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:namespace:ActionExecutionInput', actionExecutionInput_namespace - The variable namespace associated with the action. All variables produced as output by this action fall under this namespace.

$sel:resolvedConfiguration:ActionExecutionInput', actionExecutionInput_resolvedConfiguration - Configuration data for an action execution with all variable references replaced with their real values for the execution.

$sel:region:ActionExecutionInput', actionExecutionInput_region - The AWS Region for the action, such as us-east-1.

$sel:configuration:ActionExecutionInput', actionExecutionInput_configuration - Configuration data for an action execution.

$sel:actionTypeId:ActionExecutionInput', actionExecutionInput_actionTypeId - Undocumented member.

$sel:inputArtifacts:ActionExecutionInput', actionExecutionInput_inputArtifacts - Details of input artifacts of the action that correspond to the action execution.

$sel:roleArn:ActionExecutionInput', actionExecutionInput_roleArn - The ARN of the IAM service role that performs the declared action. This is assumed through the roleArn for the pipeline.

actionExecutionInput_namespace :: Lens' ActionExecutionInput (Maybe Text)

The variable namespace associated with the action. All variables produced as output by this action fall under this namespace.

actionExecutionInput_resolvedConfiguration :: Lens' ActionExecutionInput (Maybe (HashMap Text Text))

Configuration data for an action execution with all variable references replaced with their real values for the execution.

actionExecutionInput_region :: Lens' ActionExecutionInput (Maybe Text)

The AWS Region for the action, such as us-east-1.

actionExecutionInput_configuration :: Lens' ActionExecutionInput (Maybe (HashMap Text Text))

Configuration data for an action execution.

actionExecutionInput_actionTypeId :: Lens' ActionExecutionInput (Maybe ActionTypeId)

Undocumented member.

actionExecutionInput_inputArtifacts :: Lens' ActionExecutionInput (Maybe [ArtifactDetail])

Details of input artifacts of the action that correspond to the action execution.

actionExecutionInput_roleArn :: Lens' ActionExecutionInput (Maybe Text)

The ARN of the IAM service role that performs the declared action. This is assumed through the roleArn for the pipeline.