Network.AWS.CodePipeline.Types.ActionExecutionOutput

Description

Documentation

data ActionExecutionOutput

Output details listed for an action execution, such as the action execution result.

See: newActionExecutionOutput smart constructor.

Constructors

ActionExecutionOutput'

Fields

executionResult :: Maybe ActionExecutionResult

Execution result information listed in the output details for an action execution.

outputArtifacts :: Maybe [ArtifactDetail]

Details of output artifacts of the action that correspond to the action execution.

outputVariables :: Maybe (HashMap Text Text)

The outputVariables field shows the key-value pairs that were output as part of that execution.

newActionExecutionOutput :: ActionExecutionOutput

Create a value of ActionExecutionOutput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:executionResult:ActionExecutionOutput', actionExecutionOutput_executionResult - Execution result information listed in the output details for an action execution.

$sel:outputArtifacts:ActionExecutionOutput', actionExecutionOutput_outputArtifacts - Details of output artifacts of the action that correspond to the action execution.

$sel:outputVariables:ActionExecutionOutput', actionExecutionOutput_outputVariables - The outputVariables field shows the key-value pairs that were output as part of that execution.

actionExecutionOutput_executionResult :: Lens' ActionExecutionOutput (Maybe ActionExecutionResult)

Execution result information listed in the output details for an action execution.

actionExecutionOutput_outputArtifacts :: Lens' ActionExecutionOutput (Maybe [ArtifactDetail])

Details of output artifacts of the action that correspond to the action execution.

actionExecutionOutput_outputVariables :: Lens' ActionExecutionOutput (Maybe (HashMap Text Text))

The outputVariables field shows the key-value pairs that were output as part of that execution.