Network.AWS.CodePipeline.Types.ActionState

Description

Documentation

data ActionState

Represents information about the state of an action.

See: newActionState smart constructor.

Constructors

ActionState'

Fields

revisionUrl :: Maybe Text

A URL link for more information about the revision, such as a commit details page.

entityUrl :: Maybe Text

A URL link for more information about the state of the action, such as a deployment group details page.

actionName :: Maybe Text

The name of the action.

currentRevision :: Maybe ActionRevision

Represents information about the version (or revision) of an action.

latestExecution :: Maybe ActionExecution

Represents information about the run of an action.

newActionState :: ActionState

Create a value of ActionState with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:revisionUrl:ActionState', actionState_revisionUrl - A URL link for more information about the revision, such as a commit details page.

$sel:entityUrl:ActionState', actionState_entityUrl - A URL link for more information about the state of the action, such as a deployment group details page.

$sel:actionName:ActionState', actionState_actionName - The name of the action.

$sel:currentRevision:ActionState', actionState_currentRevision - Represents information about the version (or revision) of an action.

$sel:latestExecution:ActionState', actionState_latestExecution - Represents information about the run of an action.

actionState_revisionUrl :: Lens' ActionState (Maybe Text)

A URL link for more information about the revision, such as a commit details page.

actionState_entityUrl :: Lens' ActionState (Maybe Text)

A URL link for more information about the state of the action, such as a deployment group details page.

actionState_actionName :: Lens' ActionState (Maybe Text)

The name of the action.

actionState_currentRevision :: Lens' ActionState (Maybe ActionRevision)

Represents information about the version (or revision) of an action.

actionState_latestExecution :: Lens' ActionState (Maybe ActionExecution)

Represents information about the run of an action.