Network.AWS.CodePipeline.Types.ActionTypeDeclaration

Description

Documentation

data ActionTypeDeclaration

The parameters for the action type definition that are provided when the action type is created or updated.

See: newActionTypeDeclaration smart constructor.

Constructors

ActionTypeDeclaration'

Fields

urls :: Maybe ActionTypeUrls

The links associated with the action type to be updated.

permissions :: Maybe ActionTypePermissions

Details identifying the accounts with permissions to use the action type.

description :: Maybe Text

The description for the action type to be updated.

properties :: Maybe [ActionTypeProperty]

The properties of the action type to be updated.

executor :: ActionTypeExecutor

Information about the executor for an action type that was created with any supported integration model.

id :: ActionTypeIdentifier

The action category, owner, provider, and version of the action type to be updated.

inputArtifactDetails :: ActionTypeArtifactDetails

Details for the artifacts, such as application files, to be worked on by the action. For example, the minimum and maximum number of input artifacts allowed.

outputArtifactDetails :: ActionTypeArtifactDetails

Details for the output artifacts, such as a built application, that are the result of the action. For example, the minimum and maximum number of output artifacts allowed.

newActionTypeDeclaration

Arguments

:: ActionTypeExecutor
$sel:executor:ActionTypeDeclaration'

-> ActionTypeIdentifier
$sel:id:ActionTypeDeclaration'

-> ActionTypeArtifactDetails
$sel:inputArtifactDetails:ActionTypeDeclaration'

-> ActionTypeArtifactDetails
$sel:outputArtifactDetails:ActionTypeDeclaration'

-> ActionTypeDeclaration

Create a value of ActionTypeDeclaration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:urls:ActionTypeDeclaration', actionTypeDeclaration_urls - The links associated with the action type to be updated.

$sel:permissions:ActionTypeDeclaration', actionTypeDeclaration_permissions - Details identifying the accounts with permissions to use the action type.

$sel:description:ActionTypeDeclaration', actionTypeDeclaration_description - The description for the action type to be updated.

$sel:properties:ActionTypeDeclaration', actionTypeDeclaration_properties - The properties of the action type to be updated.

$sel:executor:ActionTypeDeclaration', actionTypeDeclaration_executor - Information about the executor for an action type that was created with any supported integration model.

$sel:id:ActionTypeDeclaration', actionTypeDeclaration_id - The action category, owner, provider, and version of the action type to be updated.

$sel:inputArtifactDetails:ActionTypeDeclaration', actionTypeDeclaration_inputArtifactDetails - Details for the artifacts, such as application files, to be worked on by the action. For example, the minimum and maximum number of input artifacts allowed.

$sel:outputArtifactDetails:ActionTypeDeclaration', actionTypeDeclaration_outputArtifactDetails - Details for the output artifacts, such as a built application, that are the result of the action. For example, the minimum and maximum number of output artifacts allowed.

actionTypeDeclaration_urls :: Lens' ActionTypeDeclaration (Maybe ActionTypeUrls)

The links associated with the action type to be updated.

actionTypeDeclaration_permissions :: Lens' ActionTypeDeclaration (Maybe ActionTypePermissions)

Details identifying the accounts with permissions to use the action type.

actionTypeDeclaration_description :: Lens' ActionTypeDeclaration (Maybe Text)

The description for the action type to be updated.

actionTypeDeclaration_properties :: Lens' ActionTypeDeclaration (Maybe [ActionTypeProperty])

The properties of the action type to be updated.

actionTypeDeclaration_executor :: Lens' ActionTypeDeclaration ActionTypeExecutor

Information about the executor for an action type that was created with any supported integration model.

actionTypeDeclaration_id :: Lens' ActionTypeDeclaration ActionTypeIdentifier

The action category, owner, provider, and version of the action type to be updated.

actionTypeDeclaration_inputArtifactDetails :: Lens' ActionTypeDeclaration ActionTypeArtifactDetails

Details for the artifacts, such as application files, to be worked on by the action. For example, the minimum and maximum number of input artifacts allowed.

actionTypeDeclaration_outputArtifactDetails :: Lens' ActionTypeDeclaration ActionTypeArtifactDetails

Details for the output artifacts, such as a built application, that are the result of the action. For example, the minimum and maximum number of output artifacts allowed.