Network.AWS.CodePipeline.Types.ActionTypeId

Description

Documentation

data ActionTypeId

Represents information about an action type.

See: newActionTypeId smart constructor.

Constructors

ActionTypeId'

Fields

category :: ActionCategory

A category defines what kind of action can be taken in the stage, and constrains the provider type for the action. Valid categories are limited to one of the following values.

Source

Build

Test

Deploy

Invoke

Approval

owner :: ActionOwner

The creator of the action being called. There are three valid values for the Owner field in the action category section within your pipeline structure: AWS, ThirdParty, and Custom. For more information, see Valid Action Types and Providers in CodePipeline.

provider :: Text

The provider of the service being called by the action. Valid providers are determined by the action category. For example, an action in the Deploy category type might have a provider of AWS CodeDeploy, which would be specified as CodeDeploy. For more information, see Valid Action Types and Providers in CodePipeline.

version :: Text

A string that describes the action version.

newActionTypeId

Arguments

:: ActionCategory
$sel:category:ActionTypeId'

-> ActionOwner
$sel:owner:ActionTypeId'

-> Text
$sel:provider:ActionTypeId'

-> Text
$sel:version:ActionTypeId'

-> ActionTypeId

Create a value of ActionTypeId with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:category:ActionTypeId', actionTypeId_category - A category defines what kind of action can be taken in the stage, and constrains the provider type for the action. Valid categories are limited to one of the following values.

Source

Build

Test

Deploy

Invoke

Approval

$sel:owner:ActionTypeId', actionTypeId_owner - The creator of the action being called. There are three valid values for the Owner field in the action category section within your pipeline structure: AWS, ThirdParty, and Custom. For more information, see Valid Action Types and Providers in CodePipeline.

$sel:provider:ActionTypeId', actionTypeId_provider - The provider of the service being called by the action. Valid providers are determined by the action category. For example, an action in the Deploy category type might have a provider of AWS CodeDeploy, which would be specified as CodeDeploy. For more information, see Valid Action Types and Providers in CodePipeline.

$sel:version:ActionTypeId', actionTypeId_version - A string that describes the action version.

actionTypeId_category :: Lens' ActionTypeId ActionCategory

A category defines what kind of action can be taken in the stage, and constrains the provider type for the action. Valid categories are limited to one of the following values.

Source

Build

Test

Deploy

Invoke

Approval

actionTypeId_owner :: Lens' ActionTypeId ActionOwner

The creator of the action being called. There are three valid values for the Owner field in the action category section within your pipeline structure: AWS, ThirdParty, and Custom. For more information, see Valid Action Types and Providers in CodePipeline.

actionTypeId_provider :: Lens' ActionTypeId Text

The provider of the service being called by the action. Valid providers are determined by the action category. For example, an action in the Deploy category type might have a provider of AWS CodeDeploy, which would be specified as CodeDeploy. For more information, see Valid Action Types and Providers in CodePipeline.

actionTypeId_version :: Lens' ActionTypeId Text

A string that describes the action version.