Network.AWS.CodePipeline.Types.ActionTypeProperty

Description

Documentation

data ActionTypeProperty

Represents information about each property specified in the action configuration, such as the description and key name that display for the customer using the action type.

See: newActionTypeProperty smart constructor.

Constructors

ActionTypeProperty'

Fields

queryable :: Maybe Bool

Indicates that the property is used with polling. An action type can have up to one queryable property. If it has one, that property must be both required and not secret.

description :: Maybe Text

The description of the property that is displayed to users.

name :: Text

The property name that is displayed to users.

optional :: Bool

Whether the configuration property is an optional value.

key :: Bool

Whether the configuration property is a key.

noEcho :: Bool

Whether to omit the field value entered by the customer in the log. If true, the value is not saved in CloudTrail logs for the action execution.

newActionTypeProperty

Arguments

:: Text
$sel:name:ActionTypeProperty'

-> Bool
$sel:optional:ActionTypeProperty'

-> Bool
$sel:key:ActionTypeProperty'

-> Bool
$sel:noEcho:ActionTypeProperty'

-> ActionTypeProperty

Create a value of ActionTypeProperty with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:queryable:ActionTypeProperty', actionTypeProperty_queryable - Indicates that the property is used with polling. An action type can have up to one queryable property. If it has one, that property must be both required and not secret.

$sel:description:ActionTypeProperty', actionTypeProperty_description - The description of the property that is displayed to users.

$sel:name:ActionTypeProperty', actionTypeProperty_name - The property name that is displayed to users.

$sel:optional:ActionTypeProperty', actionTypeProperty_optional - Whether the configuration property is an optional value.

$sel:key:ActionTypeProperty', actionTypeProperty_key - Whether the configuration property is a key.

$sel:noEcho:ActionTypeProperty', actionTypeProperty_noEcho - Whether to omit the field value entered by the customer in the log. If true, the value is not saved in CloudTrail logs for the action execution.

actionTypeProperty_queryable :: Lens' ActionTypeProperty (Maybe Bool)

Indicates that the property is used with polling. An action type can have up to one queryable property. If it has one, that property must be both required and not secret.

actionTypeProperty_description :: Lens' ActionTypeProperty (Maybe Text)

The description of the property that is displayed to users.

actionTypeProperty_name :: Lens' ActionTypeProperty Text

The property name that is displayed to users.

actionTypeProperty_optional :: Lens' ActionTypeProperty Bool

Whether the configuration property is an optional value.

actionTypeProperty_key :: Lens' ActionTypeProperty Bool

Whether the configuration property is a key.

actionTypeProperty_noEcho :: Lens' ActionTypeProperty Bool

Whether to omit the field value entered by the customer in the log. If true, the value is not saved in CloudTrail logs for the action execution.