Network.AWS.CodePipeline.Types.ActionTypeUrls

Description

Documentation

data ActionTypeUrls

Returns information about URLs for web pages that display to customers as links on the pipeline view, such as an external configuration page for the action type.

See: newActionTypeUrls smart constructor.

Constructors

ActionTypeUrls'

Fields

executionUrlTemplate :: Maybe Text

The link to an execution page for the action type in progress. For example, for a CodeDeploy action, this link is shown on the pipeline view page in the CodePipeline console, and it links to a CodeDeploy status page.

revisionUrlTemplate :: Maybe Text

The URL returned to the CodePipeline console that contains a link to the page where customers can update or change the configuration of the external action.

entityUrlTemplate :: Maybe Text

The URL returned to the CodePipeline console that provides a deep link to the resources of the external system, such as a status page. This link is provided as part of the action display in the pipeline.

configurationUrl :: Maybe Text

The URL returned to the CodePipeline console that contains a link to the page where customers can configure the external action.

newActionTypeUrls :: ActionTypeUrls

Create a value of ActionTypeUrls with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:executionUrlTemplate:ActionTypeUrls', actionTypeUrls_executionUrlTemplate - The link to an execution page for the action type in progress. For example, for a CodeDeploy action, this link is shown on the pipeline view page in the CodePipeline console, and it links to a CodeDeploy status page.

$sel:revisionUrlTemplate:ActionTypeUrls', actionTypeUrls_revisionUrlTemplate - The URL returned to the CodePipeline console that contains a link to the page where customers can update or change the configuration of the external action.

$sel:entityUrlTemplate:ActionTypeUrls', actionTypeUrls_entityUrlTemplate - The URL returned to the CodePipeline console that provides a deep link to the resources of the external system, such as a status page. This link is provided as part of the action display in the pipeline.

$sel:configurationUrl:ActionTypeUrls', actionTypeUrls_configurationUrl - The URL returned to the CodePipeline console that contains a link to the page where customers can configure the external action.

actionTypeUrls_executionUrlTemplate :: Lens' ActionTypeUrls (Maybe Text)

The link to an execution page for the action type in progress. For example, for a CodeDeploy action, this link is shown on the pipeline view page in the CodePipeline console, and it links to a CodeDeploy status page.

actionTypeUrls_revisionUrlTemplate :: Lens' ActionTypeUrls (Maybe Text)

The URL returned to the CodePipeline console that contains a link to the page where customers can update or change the configuration of the external action.

actionTypeUrls_entityUrlTemplate :: Lens' ActionTypeUrls (Maybe Text)

The URL returned to the CodePipeline console that provides a deep link to the resources of the external system, such as a status page. This link is provided as part of the action display in the pipeline.

actionTypeUrls_configurationUrl :: Lens' ActionTypeUrls (Maybe Text)

The URL returned to the CodePipeline console that contains a link to the page where customers can configure the external action.