Network.AWS.CodePipeline.Types.ArtifactDetail

Description

Documentation

data ArtifactDetail

Artifact details for the action execution, such as the artifact location.

See: newArtifactDetail smart constructor.

Constructors

ArtifactDetail'

Fields

s3location :: Maybe S3Location

The Amazon S3 artifact location for the action execution.

name :: Maybe Text

The artifact object name for the action execution.

newArtifactDetail :: ArtifactDetail

Create a value of ArtifactDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3location:ArtifactDetail', artifactDetail_s3location - The Amazon S3 artifact location for the action execution.

$sel:name:ArtifactDetail', artifactDetail_name - The artifact object name for the action execution.

artifactDetail_s3location :: Lens' ArtifactDetail (Maybe S3Location)

The Amazon S3 artifact location for the action execution.

artifactDetail_name :: Lens' ArtifactDetail (Maybe Text)

The artifact object name for the action execution.