Network.AWS.CodePipeline.Types.ArtifactRevision

Description

Documentation

data ArtifactRevision

Represents revision details of an artifact.

See: newArtifactRevision smart constructor.

Constructors

ArtifactRevision'

Fields

revisionSummary :: Maybe Text

Summary information about the most recent revision of the artifact. For GitHub and AWS CodeCommit repositories, the commit message. For Amazon S3 buckets or actions, the user-provided content of a codepipeline-artifact-revision-summary key specified in the object metadata.

revisionUrl :: Maybe Text

The commit ID for the artifact revision. For artifacts stored in GitHub or AWS CodeCommit repositories, the commit ID is linked to a commit details page.

created :: Maybe POSIX

The date and time when the most recent revision of the artifact was created, in timestamp format.

name :: Maybe Text

The name of an artifact. This name might be system-generated, such as "MyApp", or defined by the user when an action is created.

revisionId :: Maybe Text

The revision ID of the artifact.

revisionChangeIdentifier :: Maybe Text

An additional identifier for a revision, such as a commit date or, for artifacts stored in Amazon S3 buckets, the ETag value.

newArtifactRevision :: ArtifactRevision

Create a value of ArtifactRevision with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:revisionSummary:ArtifactRevision', artifactRevision_revisionSummary - Summary information about the most recent revision of the artifact. For GitHub and AWS CodeCommit repositories, the commit message. For Amazon S3 buckets or actions, the user-provided content of a codepipeline-artifact-revision-summary key specified in the object metadata.

$sel:revisionUrl:ArtifactRevision', artifactRevision_revisionUrl - The commit ID for the artifact revision. For artifacts stored in GitHub or AWS CodeCommit repositories, the commit ID is linked to a commit details page.

$sel:created:ArtifactRevision', artifactRevision_created - The date and time when the most recent revision of the artifact was created, in timestamp format.

$sel:name:ArtifactRevision', artifactRevision_name - The name of an artifact. This name might be system-generated, such as "MyApp", or defined by the user when an action is created.

$sel:revisionId:ArtifactRevision', artifactRevision_revisionId - The revision ID of the artifact.

$sel:revisionChangeIdentifier:ArtifactRevision', artifactRevision_revisionChangeIdentifier - An additional identifier for a revision, such as a commit date or, for artifacts stored in Amazon S3 buckets, the ETag value.

artifactRevision_revisionSummary :: Lens' ArtifactRevision (Maybe Text)

Summary information about the most recent revision of the artifact. For GitHub and AWS CodeCommit repositories, the commit message. For Amazon S3 buckets or actions, the user-provided content of a codepipeline-artifact-revision-summary key specified in the object metadata.

artifactRevision_revisionUrl :: Lens' ArtifactRevision (Maybe Text)

The commit ID for the artifact revision. For artifacts stored in GitHub or AWS CodeCommit repositories, the commit ID is linked to a commit details page.

artifactRevision_created :: Lens' ArtifactRevision (Maybe UTCTime)

The date and time when the most recent revision of the artifact was created, in timestamp format.

artifactRevision_name :: Lens' ArtifactRevision (Maybe Text)

The name of an artifact. This name might be system-generated, such as "MyApp", or defined by the user when an action is created.

artifactRevision_revisionId :: Lens' ArtifactRevision (Maybe Text)

The revision ID of the artifact.

artifactRevision_revisionChangeIdentifier :: Lens' ArtifactRevision (Maybe Text)

An additional identifier for a revision, such as a commit date or, for artifacts stored in Amazon S3 buckets, the ETag value.