Network.AWS.CodePipeline.Types.JobData

Description

Documentation

data JobData

Represents other information about a job required for a job worker to complete the job.

See: newJobData smart constructor.

Constructors

JobData'

Fields

continuationToken :: Maybe Text

A system-generated token, such as a AWS CodeDeploy deployment ID, required by a job to continue the job asynchronously.

outputArtifacts :: Maybe [Artifact]

The output of the job.

artifactCredentials :: Maybe (Sensitive AWSSessionCredentials)

Represents an AWS session credentials object. These credentials are temporary credentials that are issued by AWS Secure Token Service (STS). They can be used to access input and output artifacts in the S3 bucket used to store artifacts for the pipeline in AWS CodePipeline.

pipelineContext :: Maybe PipelineContext

Represents information about a pipeline to a job worker.

Includes pipelineArn and pipelineExecutionId for custom jobs.

encryptionKey :: Maybe EncryptionKey

Represents information about the key used to encrypt data in the artifact store, such as an AWS Key Management Service (AWS KMS) key.

actionTypeId :: Maybe ActionTypeId

Represents information about an action type.

inputArtifacts :: Maybe [Artifact]

The artifact supplied to the job.

actionConfiguration :: Maybe ActionConfiguration

Represents information about an action configuration.

newJobData :: JobData

Create a value of JobData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:continuationToken:JobData', jobData_continuationToken - A system-generated token, such as a AWS CodeDeploy deployment ID, required by a job to continue the job asynchronously.

$sel:outputArtifacts:JobData', jobData_outputArtifacts - The output of the job.

$sel:artifactCredentials:JobData', jobData_artifactCredentials - Represents an AWS session credentials object. These credentials are temporary credentials that are issued by AWS Secure Token Service (STS). They can be used to access input and output artifacts in the S3 bucket used to store artifacts for the pipeline in AWS CodePipeline.

$sel:pipelineContext:JobData', jobData_pipelineContext - Represents information about a pipeline to a job worker.

Includes pipelineArn and pipelineExecutionId for custom jobs.

$sel:encryptionKey:JobData', jobData_encryptionKey - Represents information about the key used to encrypt data in the artifact store, such as an AWS Key Management Service (AWS KMS) key.

$sel:actionTypeId:JobData', jobData_actionTypeId - Represents information about an action type.

$sel:inputArtifacts:JobData', jobData_inputArtifacts - The artifact supplied to the job.

$sel:actionConfiguration:JobData', jobData_actionConfiguration - Represents information about an action configuration.

jobData_continuationToken :: Lens' JobData (Maybe Text)

A system-generated token, such as a AWS CodeDeploy deployment ID, required by a job to continue the job asynchronously.

jobData_outputArtifacts :: Lens' JobData (Maybe [Artifact])

The output of the job.

jobData_artifactCredentials :: Lens' JobData (Maybe AWSSessionCredentials)

Represents an AWS session credentials object. These credentials are temporary credentials that are issued by AWS Secure Token Service (STS). They can be used to access input and output artifacts in the S3 bucket used to store artifacts for the pipeline in AWS CodePipeline.

jobData_pipelineContext :: Lens' JobData (Maybe PipelineContext)

Represents information about a pipeline to a job worker.

Includes pipelineArn and pipelineExecutionId for custom jobs.

jobData_encryptionKey :: Lens' JobData (Maybe EncryptionKey)

Represents information about the key used to encrypt data in the artifact store, such as an AWS Key Management Service (AWS KMS) key.

jobData_actionTypeId :: Lens' JobData (Maybe ActionTypeId)

Represents information about an action type.

jobData_inputArtifacts :: Lens' JobData (Maybe [Artifact])

The artifact supplied to the job.

jobData_actionConfiguration :: Lens' JobData (Maybe ActionConfiguration)

Represents information about an action configuration.