Network.AWS.CodePipeline.Types.JobDetails

Description

Documentation

data JobDetails

Represents information about the details of a job.

See: newJobDetails smart constructor.

Constructors

JobDetails'

Fields

data' :: Maybe JobData

Represents other information about a job required for a job worker to complete the job.

accountId :: Maybe Text

The AWS account ID associated with the job.

id :: Maybe Text

The unique system-generated ID of the job.

newJobDetails :: JobDetails

Create a value of JobDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:data':JobDetails', jobDetails_data - Represents other information about a job required for a job worker to complete the job.

$sel:accountId:JobDetails', jobDetails_accountId - The AWS account ID associated with the job.

$sel:id:JobDetails', jobDetails_id - The unique system-generated ID of the job.

jobDetails_data :: Lens' JobDetails (Maybe JobData)

Represents other information about a job required for a job worker to complete the job.

jobDetails_accountId :: Lens' JobDetails (Maybe Text)

The AWS account ID associated with the job.

jobDetails_id :: Lens' JobDetails (Maybe Text)

The unique system-generated ID of the job.