Network.AWS.CodePipeline.Types.ListWebhookItem

Description

Documentation

data ListWebhookItem

The detail returned for each webhook after listing webhooks, such as the webhook URL, the webhook name, and the webhook ARN.

See: newListWebhookItem smart constructor.

Constructors

ListWebhookItem'

Fields

lastTriggered :: Maybe POSIX

The date and time a webhook was last successfully triggered, in timestamp format.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the webhook.

tags :: Maybe [Tag]

Specifies the tags applied to the webhook.

errorMessage :: Maybe Text

The text of the error message about the webhook.

errorCode :: Maybe Text

The number code of the error.

definition :: WebhookDefinition

The detail returned for each webhook, such as the webhook authentication type and filter rules.

url :: Text

A unique URL generated by CodePipeline. When a POST request is made to this URL, the defined pipeline is started as long as the body of the post request satisfies the defined authentication and filtering conditions. Deleting and re-creating a webhook makes the old URL invalid and generates a new one.

newListWebhookItem

Arguments

:: WebhookDefinition
$sel:definition:ListWebhookItem'

-> Text
$sel:url:ListWebhookItem'

-> ListWebhookItem

Create a value of ListWebhookItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastTriggered:ListWebhookItem', listWebhookItem_lastTriggered - The date and time a webhook was last successfully triggered, in timestamp format.

$sel:arn:ListWebhookItem', listWebhookItem_arn - The Amazon Resource Name (ARN) of the webhook.

$sel:tags:ListWebhookItem', listWebhookItem_tags - Specifies the tags applied to the webhook.

$sel:errorMessage:ListWebhookItem', listWebhookItem_errorMessage - The text of the error message about the webhook.

$sel:errorCode:ListWebhookItem', listWebhookItem_errorCode - The number code of the error.

$sel:definition:ListWebhookItem', listWebhookItem_definition - The detail returned for each webhook, such as the webhook authentication type and filter rules.

$sel:url:ListWebhookItem', listWebhookItem_url - A unique URL generated by CodePipeline. When a POST request is made to this URL, the defined pipeline is started as long as the body of the post request satisfies the defined authentication and filtering conditions. Deleting and re-creating a webhook makes the old URL invalid and generates a new one.

listWebhookItem_lastTriggered :: Lens' ListWebhookItem (Maybe UTCTime)

The date and time a webhook was last successfully triggered, in timestamp format.

listWebhookItem_arn :: Lens' ListWebhookItem (Maybe Text)

The Amazon Resource Name (ARN) of the webhook.

listWebhookItem_tags :: Lens' ListWebhookItem (Maybe [Tag])

Specifies the tags applied to the webhook.

listWebhookItem_errorMessage :: Lens' ListWebhookItem (Maybe Text)

The text of the error message about the webhook.

listWebhookItem_errorCode :: Lens' ListWebhookItem (Maybe Text)

The number code of the error.

listWebhookItem_definition :: Lens' ListWebhookItem WebhookDefinition

The detail returned for each webhook, such as the webhook authentication type and filter rules.

listWebhookItem_url :: Lens' ListWebhookItem Text

A unique URL generated by CodePipeline. When a POST request is made to this URL, the defined pipeline is started as long as the body of the post request satisfies the defined authentication and filtering conditions. Deleting and re-creating a webhook makes the old URL invalid and generates a new one.