Network.AWS.CodePipeline.Types.PipelineDeclaration

Description

Documentation

data PipelineDeclaration

Represents the structure of actions and stages to be performed in the pipeline.

See: newPipelineDeclaration smart constructor.

Constructors

PipelineDeclaration'

Fields

version :: Maybe Natural

The version number of the pipeline. A new pipeline always has a version number of 1. This number is incremented when a pipeline is updated.

artifactStores :: Maybe (HashMap Text ArtifactStore)

A mapping of artifactStore objects and their corresponding AWS Regions. There must be an artifact store for the pipeline Region and for each cross-region action in the pipeline.

You must include either artifactStore or artifactStores in your pipeline, but you cannot use both. If you create a cross-region action in your pipeline, you must use artifactStores.

artifactStore :: Maybe ArtifactStore

Represents information about the S3 bucket where artifacts are stored for the pipeline.

You must include either artifactStore or artifactStores in your pipeline, but you cannot use both. If you create a cross-region action in your pipeline, you must use artifactStores.

name :: Text

The name of the pipeline.

roleArn :: Text

The Amazon Resource Name (ARN) for AWS CodePipeline to use to either perform actions with no actionRoleArn, or to use to assume roles for actions with an actionRoleArn.

stages :: [StageDeclaration]

The stage in which to perform the action.

newPipelineDeclaration

Arguments

:: Text
$sel:name:PipelineDeclaration'

-> Text
$sel:roleArn:PipelineDeclaration'

-> PipelineDeclaration

Create a value of PipelineDeclaration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:version:PipelineDeclaration', pipelineDeclaration_version - The version number of the pipeline. A new pipeline always has a version number of 1. This number is incremented when a pipeline is updated.

$sel:artifactStores:PipelineDeclaration', pipelineDeclaration_artifactStores - A mapping of artifactStore objects and their corresponding AWS Regions. There must be an artifact store for the pipeline Region and for each cross-region action in the pipeline.

You must include either artifactStore or artifactStores in your pipeline, but you cannot use both. If you create a cross-region action in your pipeline, you must use artifactStores.

$sel:artifactStore:PipelineDeclaration', pipelineDeclaration_artifactStore - Represents information about the S3 bucket where artifacts are stored for the pipeline.

You must include either artifactStore or artifactStores in your pipeline, but you cannot use both. If you create a cross-region action in your pipeline, you must use artifactStores.

$sel:name:PipelineDeclaration', pipelineDeclaration_name - The name of the pipeline.

$sel:roleArn:PipelineDeclaration', pipelineDeclaration_roleArn - The Amazon Resource Name (ARN) for AWS CodePipeline to use to either perform actions with no actionRoleArn, or to use to assume roles for actions with an actionRoleArn.

$sel:stages:PipelineDeclaration', pipelineDeclaration_stages - The stage in which to perform the action.

pipelineDeclaration_version :: Lens' PipelineDeclaration (Maybe Natural)

The version number of the pipeline. A new pipeline always has a version number of 1. This number is incremented when a pipeline is updated.

pipelineDeclaration_artifactStores :: Lens' PipelineDeclaration (Maybe (HashMap Text ArtifactStore))

A mapping of artifactStore objects and their corresponding AWS Regions. There must be an artifact store for the pipeline Region and for each cross-region action in the pipeline.

You must include either artifactStore or artifactStores in your pipeline, but you cannot use both. If you create a cross-region action in your pipeline, you must use artifactStores.

pipelineDeclaration_artifactStore :: Lens' PipelineDeclaration (Maybe ArtifactStore)

Represents information about the S3 bucket where artifacts are stored for the pipeline.

You must include either artifactStore or artifactStores in your pipeline, but you cannot use both. If you create a cross-region action in your pipeline, you must use artifactStores.

pipelineDeclaration_name :: Lens' PipelineDeclaration Text

The name of the pipeline.

pipelineDeclaration_roleArn :: Lens' PipelineDeclaration Text

The Amazon Resource Name (ARN) for AWS CodePipeline to use to either perform actions with no actionRoleArn, or to use to assume roles for actions with an actionRoleArn.

pipelineDeclaration_stages :: Lens' PipelineDeclaration [StageDeclaration]

The stage in which to perform the action.