Network.AWS.CodePipeline.Types.PipelineExecution

Description

Documentation

data PipelineExecution

Represents information about an execution of a pipeline.

See: newPipelineExecution smart constructor.

Constructors

PipelineExecution'

Fields

status :: Maybe PipelineExecutionStatus

The status of the pipeline execution.

Cancelled: The pipeline’s definition was updated before the pipeline execution could be completed.

InProgress: The pipeline execution is currently running.

Stopped: The pipeline execution was manually stopped. For more information, see Stopped Executions.

Stopping: The pipeline execution received a request to be manually stopped. Depending on the selected stop mode, the execution is either completing or abandoning in-progress actions. For more information, see Stopped Executions.

Succeeded: The pipeline execution was completed successfully.

Superseded: While this pipeline execution was waiting for the next stage to be completed, a newer pipeline execution advanced and continued through the pipeline instead. For more information, see Superseded Executions.

Failed: The pipeline execution was not completed successfully.

artifactRevisions :: Maybe [ArtifactRevision]

A list of ArtifactRevision objects included in a pipeline execution.

pipelineVersion :: Maybe Natural

The version number of the pipeline with the specified pipeline execution.

statusSummary :: Maybe Text

A summary that contains a description of the pipeline execution status.

pipelineName :: Maybe Text

The name of the pipeline with the specified pipeline execution.

pipelineExecutionId :: Maybe Text

The ID of the pipeline execution.

newPipelineExecution :: PipelineExecution

Create a value of PipelineExecution with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:PipelineExecution', pipelineExecution_status - The status of the pipeline execution.

Cancelled: The pipeline’s definition was updated before the pipeline execution could be completed.

InProgress: The pipeline execution is currently running.

Stopped: The pipeline execution was manually stopped. For more information, see Stopped Executions.

Stopping: The pipeline execution received a request to be manually stopped. Depending on the selected stop mode, the execution is either completing or abandoning in-progress actions. For more information, see Stopped Executions.

Succeeded: The pipeline execution was completed successfully.

Superseded: While this pipeline execution was waiting for the next stage to be completed, a newer pipeline execution advanced and continued through the pipeline instead. For more information, see Superseded Executions.

Failed: The pipeline execution was not completed successfully.

$sel:artifactRevisions:PipelineExecution', pipelineExecution_artifactRevisions - A list of ArtifactRevision objects included in a pipeline execution.

$sel:pipelineVersion:PipelineExecution', pipelineExecution_pipelineVersion - The version number of the pipeline with the specified pipeline execution.

$sel:statusSummary:PipelineExecution', pipelineExecution_statusSummary - A summary that contains a description of the pipeline execution status.

$sel:pipelineName:PipelineExecution', pipelineExecution_pipelineName - The name of the pipeline with the specified pipeline execution.

$sel:pipelineExecutionId:PipelineExecution', pipelineExecution_pipelineExecutionId - The ID of the pipeline execution.

pipelineExecution_status :: Lens' PipelineExecution (Maybe PipelineExecutionStatus)

The status of the pipeline execution.

Cancelled: The pipeline’s definition was updated before the pipeline execution could be completed.

InProgress: The pipeline execution is currently running.

Stopped: The pipeline execution was manually stopped. For more information, see Stopped Executions.

Stopping: The pipeline execution received a request to be manually stopped. Depending on the selected stop mode, the execution is either completing or abandoning in-progress actions. For more information, see Stopped Executions.

Succeeded: The pipeline execution was completed successfully.

Superseded: While this pipeline execution was waiting for the next stage to be completed, a newer pipeline execution advanced and continued through the pipeline instead. For more information, see Superseded Executions.

Failed: The pipeline execution was not completed successfully.

pipelineExecution_artifactRevisions :: Lens' PipelineExecution (Maybe [ArtifactRevision])

A list of ArtifactRevision objects included in a pipeline execution.

pipelineExecution_pipelineVersion :: Lens' PipelineExecution (Maybe Natural)

The version number of the pipeline with the specified pipeline execution.

pipelineExecution_statusSummary :: Lens' PipelineExecution (Maybe Text)

A summary that contains a description of the pipeline execution status.

pipelineExecution_pipelineName :: Lens' PipelineExecution (Maybe Text)

The name of the pipeline with the specified pipeline execution.

pipelineExecution_pipelineExecutionId :: Lens' PipelineExecution (Maybe Text)

The ID of the pipeline execution.