Network.AWS.CodePipeline.Types.PipelineExecutionSummary

Description

Documentation

data PipelineExecutionSummary

Summary information about a pipeline execution.

See: newPipelineExecutionSummary smart constructor.

Constructors

PipelineExecutionSummary'

Fields

status :: Maybe PipelineExecutionStatus

The status of the pipeline execution.

InProgress: The pipeline execution is currently running.

Stopped: The pipeline execution was manually stopped. For more information, see Stopped Executions.

Stopping: The pipeline execution received a request to be manually stopped. Depending on the selected stop mode, the execution is either completing or abandoning in-progress actions. For more information, see Stopped Executions.

Succeeded: The pipeline execution was completed successfully.

Superseded: While this pipeline execution was waiting for the next stage to be completed, a newer pipeline execution advanced and continued through the pipeline instead. For more information, see Superseded Executions.

Failed: The pipeline execution was not completed successfully.

lastUpdateTime :: Maybe POSIX

The date and time of the last change to the pipeline execution, in timestamp format.

trigger :: Maybe ExecutionTrigger

The interaction or event that started a pipeline execution, such as automated change detection or a StartPipelineExecution API call.

startTime :: Maybe POSIX

The date and time when the pipeline execution began, in timestamp format.

stopTrigger :: Maybe StopExecutionTrigger

The interaction that stopped a pipeline execution.

sourceRevisions :: Maybe [SourceRevision]

A list of the source artifact revisions that initiated a pipeline execution.

pipelineExecutionId :: Maybe Text

The ID of the pipeline execution.

newPipelineExecutionSummary :: PipelineExecutionSummary

Create a value of PipelineExecutionSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:PipelineExecutionSummary', pipelineExecutionSummary_status - The status of the pipeline execution.

InProgress: The pipeline execution is currently running.

Stopped: The pipeline execution was manually stopped. For more information, see Stopped Executions.

Stopping: The pipeline execution received a request to be manually stopped. Depending on the selected stop mode, the execution is either completing or abandoning in-progress actions. For more information, see Stopped Executions.

Succeeded: The pipeline execution was completed successfully.

Superseded: While this pipeline execution was waiting for the next stage to be completed, a newer pipeline execution advanced and continued through the pipeline instead. For more information, see Superseded Executions.

Failed: The pipeline execution was not completed successfully.

$sel:lastUpdateTime:PipelineExecutionSummary', pipelineExecutionSummary_lastUpdateTime - The date and time of the last change to the pipeline execution, in timestamp format.

$sel:trigger:PipelineExecutionSummary', pipelineExecutionSummary_trigger - The interaction or event that started a pipeline execution, such as automated change detection or a StartPipelineExecution API call.

$sel:startTime:PipelineExecutionSummary', pipelineExecutionSummary_startTime - The date and time when the pipeline execution began, in timestamp format.

$sel:stopTrigger:PipelineExecutionSummary', pipelineExecutionSummary_stopTrigger - The interaction that stopped a pipeline execution.

$sel:sourceRevisions:PipelineExecutionSummary', pipelineExecutionSummary_sourceRevisions - A list of the source artifact revisions that initiated a pipeline execution.

$sel:pipelineExecutionId:PipelineExecutionSummary', pipelineExecutionSummary_pipelineExecutionId - The ID of the pipeline execution.

pipelineExecutionSummary_status :: Lens' PipelineExecutionSummary (Maybe PipelineExecutionStatus)

The status of the pipeline execution.

InProgress: The pipeline execution is currently running.

Stopped: The pipeline execution was manually stopped. For more information, see Stopped Executions.

Stopping: The pipeline execution received a request to be manually stopped. Depending on the selected stop mode, the execution is either completing or abandoning in-progress actions. For more information, see Stopped Executions.

Succeeded: The pipeline execution was completed successfully.

Superseded: While this pipeline execution was waiting for the next stage to be completed, a newer pipeline execution advanced and continued through the pipeline instead. For more information, see Superseded Executions.

Failed: The pipeline execution was not completed successfully.

pipelineExecutionSummary_lastUpdateTime :: Lens' PipelineExecutionSummary (Maybe UTCTime)

The date and time of the last change to the pipeline execution, in timestamp format.

pipelineExecutionSummary_trigger :: Lens' PipelineExecutionSummary (Maybe ExecutionTrigger)

The interaction or event that started a pipeline execution, such as automated change detection or a StartPipelineExecution API call.

pipelineExecutionSummary_startTime :: Lens' PipelineExecutionSummary (Maybe UTCTime)

The date and time when the pipeline execution began, in timestamp format.

pipelineExecutionSummary_stopTrigger :: Lens' PipelineExecutionSummary (Maybe StopExecutionTrigger)

The interaction that stopped a pipeline execution.

pipelineExecutionSummary_sourceRevisions :: Lens' PipelineExecutionSummary (Maybe [SourceRevision])

A list of the source artifact revisions that initiated a pipeline execution.

pipelineExecutionSummary_pipelineExecutionId :: Lens' PipelineExecutionSummary (Maybe Text)

The ID of the pipeline execution.