Network.AWS.CodePipeline.Types.ThirdPartyJobData

Description

Documentation

data ThirdPartyJobData

Represents information about the job data for a partner action.

See: newThirdPartyJobData smart constructor.

Constructors

ThirdPartyJobData'

Fields

artifactCredentials :: Maybe (Sensitive AWSSessionCredentials)

Represents an AWS session credentials object. These credentials are temporary credentials that are issued by AWS Secure Token Service (STS). They can be used to access input and output artifacts in the S3 bucket used to store artifact for the pipeline in AWS CodePipeline.

encryptionKey :: Maybe EncryptionKey

The encryption key used to encrypt and decrypt data in the artifact store for the pipeline, such as an AWS Key Management Service (AWS KMS) key. This is optional and might not be present.

actionConfiguration :: Maybe ActionConfiguration

Represents information about an action configuration.

inputArtifacts :: Maybe [Artifact]

The name of the artifact that is worked on by the action, if any. This name might be system-generated, such as "MyApp", or it might be defined by the user when the action is created. The input artifact name must match the name of an output artifact generated by an action in an earlier action or stage of the pipeline.

actionTypeId :: Maybe ActionTypeId

Represents information about an action type.

pipelineContext :: Maybe PipelineContext

Represents information about a pipeline to a job worker.

Does not include pipelineArn and pipelineExecutionId for ThirdParty jobs.

outputArtifacts :: Maybe [Artifact]

The name of the artifact that is the result of the action, if any. This name might be system-generated, such as "MyBuiltApp", or it might be defined by the user when the action is created.

continuationToken :: Maybe Text

A system-generated token, such as a AWS CodeDeploy deployment ID, that a job requires to continue the job asynchronously.

newThirdPartyJobData :: ThirdPartyJobData

Create a value of ThirdPartyJobData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:artifactCredentials:ThirdPartyJobData', thirdPartyJobData_artifactCredentials - Represents an AWS session credentials object. These credentials are temporary credentials that are issued by AWS Secure Token Service (STS). They can be used to access input and output artifacts in the S3 bucket used to store artifact for the pipeline in AWS CodePipeline.

$sel:encryptionKey:ThirdPartyJobData', thirdPartyJobData_encryptionKey - The encryption key used to encrypt and decrypt data in the artifact store for the pipeline, such as an AWS Key Management Service (AWS KMS) key. This is optional and might not be present.

$sel:actionConfiguration:ThirdPartyJobData', thirdPartyJobData_actionConfiguration - Represents information about an action configuration.

$sel:inputArtifacts:ThirdPartyJobData', thirdPartyJobData_inputArtifacts - The name of the artifact that is worked on by the action, if any. This name might be system-generated, such as "MyApp", or it might be defined by the user when the action is created. The input artifact name must match the name of an output artifact generated by an action in an earlier action or stage of the pipeline.

$sel:actionTypeId:ThirdPartyJobData', thirdPartyJobData_actionTypeId - Represents information about an action type.

$sel:pipelineContext:ThirdPartyJobData', thirdPartyJobData_pipelineContext - Represents information about a pipeline to a job worker.

Does not include pipelineArn and pipelineExecutionId for ThirdParty jobs.

$sel:outputArtifacts:ThirdPartyJobData', thirdPartyJobData_outputArtifacts - The name of the artifact that is the result of the action, if any. This name might be system-generated, such as "MyBuiltApp", or it might be defined by the user when the action is created.

$sel:continuationToken:ThirdPartyJobData', thirdPartyJobData_continuationToken - A system-generated token, such as a AWS CodeDeploy deployment ID, that a job requires to continue the job asynchronously.

thirdPartyJobData_artifactCredentials :: Lens' ThirdPartyJobData (Maybe AWSSessionCredentials)

Represents an AWS session credentials object. These credentials are temporary credentials that are issued by AWS Secure Token Service (STS). They can be used to access input and output artifacts in the S3 bucket used to store artifact for the pipeline in AWS CodePipeline.

thirdPartyJobData_encryptionKey :: Lens' ThirdPartyJobData (Maybe EncryptionKey)

The encryption key used to encrypt and decrypt data in the artifact store for the pipeline, such as an AWS Key Management Service (AWS KMS) key. This is optional and might not be present.

thirdPartyJobData_actionConfiguration :: Lens' ThirdPartyJobData (Maybe ActionConfiguration)

Represents information about an action configuration.

thirdPartyJobData_inputArtifacts :: Lens' ThirdPartyJobData (Maybe [Artifact])

The name of the artifact that is worked on by the action, if any. This name might be system-generated, such as "MyApp", or it might be defined by the user when the action is created. The input artifact name must match the name of an output artifact generated by an action in an earlier action or stage of the pipeline.

thirdPartyJobData_actionTypeId :: Lens' ThirdPartyJobData (Maybe ActionTypeId)

Represents information about an action type.

thirdPartyJobData_pipelineContext :: Lens' ThirdPartyJobData (Maybe PipelineContext)

Represents information about a pipeline to a job worker.

Does not include pipelineArn and pipelineExecutionId for ThirdParty jobs.

thirdPartyJobData_outputArtifacts :: Lens' ThirdPartyJobData (Maybe [Artifact])

The name of the artifact that is the result of the action, if any. This name might be system-generated, such as "MyBuiltApp", or it might be defined by the user when the action is created.

thirdPartyJobData_continuationToken :: Lens' ThirdPartyJobData (Maybe Text)

A system-generated token, such as a AWS CodeDeploy deployment ID, that a job requires to continue the job asynchronously.