Network.AWS.CodePipeline.Types.TransitionState

Description

Documentation

data TransitionState

Represents information about the state of transitions between one stage and another stage.

See: newTransitionState smart constructor.

Constructors

TransitionState'

Fields

enabled :: Maybe Bool

Whether the transition between stages is enabled (true) or disabled (false).

disabledReason :: Maybe Text

The user-specified reason why the transition between two stages of a pipeline was disabled.

lastChangedAt :: Maybe POSIX

The timestamp when the transition state was last changed.

lastChangedBy :: Maybe Text

The ID of the user who last changed the transition state.

newTransitionState :: TransitionState

Create a value of TransitionState with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:enabled:TransitionState', transitionState_enabled - Whether the transition between stages is enabled (true) or disabled (false).

$sel:disabledReason:TransitionState', transitionState_disabledReason - The user-specified reason why the transition between two stages of a pipeline was disabled.

$sel:lastChangedAt:TransitionState', transitionState_lastChangedAt - The timestamp when the transition state was last changed.

$sel:lastChangedBy:TransitionState', transitionState_lastChangedBy - The ID of the user who last changed the transition state.

transitionState_enabled :: Lens' TransitionState (Maybe Bool)

Whether the transition between stages is enabled (true) or disabled (false).

transitionState_disabledReason :: Lens' TransitionState (Maybe Text)

The user-specified reason why the transition between two stages of a pipeline was disabled.

transitionState_lastChangedAt :: Lens' TransitionState (Maybe UTCTime)

The timestamp when the transition state was last changed.

transitionState_lastChangedBy :: Lens' TransitionState (Maybe Text)

The ID of the user who last changed the transition state.