Network.AWS.CodePipeline.UpdateActionType

Description

Updates an action type that was created with any supported integration model, where the action type is to be used by customers of the action type provider. Use a JSON file with the action definition and UpdateActionType to provide the full structure.

Creating a Request

data UpdateActionType

See: newUpdateActionType smart constructor.

Constructors

UpdateActionType'

Fields

actionType :: ActionTypeDeclaration

The action type definition for the action type to be updated.

newUpdateActionType

Arguments

:: ActionTypeDeclaration
$sel:actionType:UpdateActionType'

-> UpdateActionType

Create a value of UpdateActionType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:actionType:UpdateActionType', updateActionType_actionType - The action type definition for the action type to be updated.

Request Lenses

updateActionType_actionType :: Lens' UpdateActionType ActionTypeDeclaration

The action type definition for the action type to be updated.

Destructuring the Response

data UpdateActionTypeResponse

See: newUpdateActionTypeResponse smart constructor.

Constructors

UpdateActionTypeResponse'

newUpdateActionTypeResponse :: UpdateActionTypeResponse

Create a value of UpdateActionTypeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.