Network.AWS.CodeStar

Description

Derived from API version 2017-04-19 of the AWS service descriptions, licensed under Apache 2.0.

AWS CodeStar

This is the API reference for AWS CodeStar. This reference provides descriptions of the operations and data types for the AWS CodeStar API along with usage examples.

You can use the AWS CodeStar API to work with:

Projects and their resources, by calling the following:

DeleteProject, which deletes a project.

DescribeProject, which lists the attributes of a project.

ListProjects, which lists all projects associated with your AWS account.

ListResources, which lists the resources associated with a project.

ListTagsForProject, which lists the tags associated with a project.

TagProject, which adds tags to a project.

UntagProject, which removes tags from a project.

UpdateProject, which updates the attributes of a project.

Teams and team members, by calling the following:

AssociateTeamMember, which adds an IAM user to the team for a project.

DisassociateTeamMember, which removes an IAM user from the team for a project.

ListTeamMembers, which lists all the IAM users in the team for a project, including their roles and attributes.

UpdateTeamMember, which updates a team member's attributes in a project.

Users, by calling the following:

CreateUserProfile, which creates a user profile that contains data associated with the user across all projects.

DeleteUserProfile, which deletes all user profile information across all projects.

DescribeUserProfile, which describes the profile of a user.

ListUserProfiles, which lists all user profiles.

UpdateUserProfile, which updates the profile for a user.

Service Configuration

defaultService :: Service

API version 2017-04-19 of the Amazon CodeStar SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by CodeStar.

TeamMemberAlreadyAssociatedException

_TeamMemberAlreadyAssociatedException :: AsError a => Getting (First ServiceError) a ServiceError

The team member is already associated with a role in this project.

ValidationException

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The specified input is either not valid, or it could not be validated.

InvalidServiceRoleException

_InvalidServiceRoleException :: AsError a => Getting (First ServiceError) a ServiceError

The service role is not valid.

ProjectCreationFailedException

_ProjectCreationFailedException :: AsError a => Getting (First ServiceError) a ServiceError

The project creation request was valid, but a nonspecific exception or error occurred during project creation. The project could not be created in AWS CodeStar.

UserProfileAlreadyExistsException

_UserProfileAlreadyExistsException :: AsError a => Getting (First ServiceError) a ServiceError

A user profile with that name already exists in this region for the AWS account. AWS CodeStar user profile names must be unique within a region for the AWS account.

ProjectNotFoundException

_ProjectNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified AWS CodeStar project was not found.

TeamMemberNotFoundException

_TeamMemberNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified team member was not found.

ProjectAlreadyExistsException

_ProjectAlreadyExistsException :: AsError a => Getting (First ServiceError) a ServiceError

An AWS CodeStar project with the same ID already exists in this region for the AWS account. AWS CodeStar project IDs must be unique within a region for the AWS account.

ProjectConfigurationException

_ProjectConfigurationException :: AsError a => Getting (First ServiceError) a ServiceError

Project configuration information is required but not specified.

ConcurrentModificationException

_ConcurrentModificationException :: AsError a => Getting (First ServiceError) a ServiceError

Another modification is being made. That modification must complete before you can make your change.

InvalidNextTokenException

_InvalidNextTokenException :: AsError a => Getting (First ServiceError) a ServiceError

The next token is not valid.

UserProfileNotFoundException

_UserProfileNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The user profile was not found.

LimitExceededException

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

A resource limit has been exceeded.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

ListProjects (Paginated)

data ListProjects

See: newListProjects smart constructor.

Constructors

ListProjects' (Maybe Text) (Maybe Natural)

newListProjects :: ListProjects

Create a value of ListProjects with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListProjects', listProjects_nextToken - The continuation token to be used to return the next set of results, if the results cannot be returned in one response.

$sel:maxResults:ListProjects', listProjects_maxResults - The maximum amount of data that can be contained in a single set of results.

data ListProjectsResponse

See: newListProjectsResponse smart constructor.

Constructors

ListProjectsResponse' (Maybe Text) Int [ProjectSummary]

newListProjectsResponse

Arguments

:: Int
$sel:httpStatus:ListProjectsResponse'

-> ListProjectsResponse

Create a value of ListProjectsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListProjects', listProjectsResponse_nextToken - The continuation token to use when requesting the next set of results, if there are more results to be returned.

$sel:httpStatus:ListProjectsResponse', listProjectsResponse_httpStatus - The response's http status code.

$sel:projects:ListProjectsResponse', listProjectsResponse_projects - A list of projects.

ListTeamMembers (Paginated)

data ListTeamMembers

See: newListTeamMembers smart constructor.

Constructors

ListTeamMembers' (Maybe Text) (Maybe Natural) Text

newListTeamMembers

Arguments

:: Text
$sel:projectId:ListTeamMembers'

-> ListTeamMembers

Create a value of ListTeamMembers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListTeamMembers', listTeamMembers_nextToken - The continuation token for the next set of results, if the results cannot be returned in one response.

$sel:maxResults:ListTeamMembers', listTeamMembers_maxResults - The maximum number of team members you want returned in a response.

$sel:projectId:ListTeamMembers', listTeamMembers_projectId - The ID of the project for which you want to list team members.

data ListTeamMembersResponse

See: newListTeamMembersResponse smart constructor.

Constructors

ListTeamMembersResponse' (Maybe Text) Int [TeamMember]

newListTeamMembersResponse

Arguments

:: Int
$sel:httpStatus:ListTeamMembersResponse'

-> ListTeamMembersResponse

Create a value of ListTeamMembersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListTeamMembers', listTeamMembersResponse_nextToken - The continuation token to use when requesting the next set of results, if there are more results to be returned.

$sel:httpStatus:ListTeamMembersResponse', listTeamMembersResponse_httpStatus - The response's http status code.

$sel:teamMembers:ListTeamMembersResponse', listTeamMembersResponse_teamMembers - A list of team member objects for the project.

DeleteProject

data DeleteProject

See: newDeleteProject smart constructor.

Constructors

DeleteProject' (Maybe Bool) (Maybe Text) Text

newDeleteProject

Arguments

:: Text
$sel:id:DeleteProject'

-> DeleteProject

Create a value of DeleteProject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deleteStack:DeleteProject', deleteProject_deleteStack - Whether to send a delete request for the primary stack in AWS CloudFormation originally used to generate the project and its resources. This option will delete all AWS resources for the project (except for any buckets in Amazon S3) as well as deleting the project itself. Recommended for most use cases.

$sel:clientRequestToken:DeleteProject', deleteProject_clientRequestToken - A user- or system-generated token that identifies the entity that requested project deletion. This token can be used to repeat the request.

$sel:id:DeleteProject', deleteProject_id - The ID of the project to be deleted in AWS CodeStar.

data DeleteProjectResponse

See: newDeleteProjectResponse smart constructor.

Constructors

DeleteProjectResponse' (Maybe Text) (Maybe Text) Int

newDeleteProjectResponse

Arguments

:: Int
$sel:httpStatus:DeleteProjectResponse'

-> DeleteProjectResponse

Create a value of DeleteProjectResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:projectArn:DeleteProjectResponse', deleteProjectResponse_projectArn - The Amazon Resource Name (ARN) of the deleted project.

$sel:stackId:DeleteProjectResponse', deleteProjectResponse_stackId - The ID of the primary stack in AWS CloudFormation that will be deleted as part of deleting the project and its resources.

$sel:httpStatus:DeleteProjectResponse', deleteProjectResponse_httpStatus - The response's http status code.

UpdateProject

data UpdateProject

See: newUpdateProject smart constructor.

Constructors

UpdateProject' (Maybe (Sensitive Text)) (Maybe (Sensitive Text)) Text

newUpdateProject

Arguments

:: Text
$sel:id:UpdateProject'

-> UpdateProject

Create a value of UpdateProject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:UpdateProject', updateProject_name - The name of the project you want to update.

$sel:description:UpdateProject', updateProject_description - The description of the project, if any.

$sel:id:UpdateProject', updateProject_id - The ID of the project you want to update.

data UpdateProjectResponse

See: newUpdateProjectResponse smart constructor.

Constructors

UpdateProjectResponse' Int

newUpdateProjectResponse

Arguments

:: Int
$sel:httpStatus:UpdateProjectResponse'

-> UpdateProjectResponse

Create a value of UpdateProjectResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateProjectResponse', updateProjectResponse_httpStatus - The response's http status code.

DisassociateTeamMember

data DisassociateTeamMember

See: newDisassociateTeamMember smart constructor.

Constructors

DisassociateTeamMember' Text Text

newDisassociateTeamMember

Arguments

:: Text
$sel:projectId:DisassociateTeamMember'

-> Text
$sel:userArn:DisassociateTeamMember'

-> DisassociateTeamMember

Create a value of DisassociateTeamMember with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:projectId:DisassociateTeamMember', disassociateTeamMember_projectId - The ID of the AWS CodeStar project from which you want to remove a team member.

$sel:userArn:DisassociateTeamMember', disassociateTeamMember_userArn - The Amazon Resource Name (ARN) of the IAM user or group whom you want to remove from the project.

data DisassociateTeamMemberResponse

See: newDisassociateTeamMemberResponse smart constructor.

Constructors

DisassociateTeamMemberResponse' Int

newDisassociateTeamMemberResponse

Arguments

:: Int
$sel:httpStatus:DisassociateTeamMemberResponse'

-> DisassociateTeamMemberResponse

Create a value of DisassociateTeamMemberResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DisassociateTeamMemberResponse', disassociateTeamMemberResponse_httpStatus - The response's http status code.

TagProject

data TagProject

See: newTagProject smart constructor.

Constructors

TagProject' Text (HashMap Text Text)

newTagProject

Arguments

:: Text
$sel:id:TagProject'

-> TagProject

Create a value of TagProject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:TagProject', tagProject_id - The ID of the project you want to add a tag to.

$sel:tags:TagProject', tagProject_tags - The tags you want to add to the project.

data TagProjectResponse

See: newTagProjectResponse smart constructor.

Constructors

TagProjectResponse' (Maybe (HashMap Text Text)) Int

newTagProjectResponse

Arguments

:: Int
$sel:httpStatus:TagProjectResponse'

-> TagProjectResponse

Create a value of TagProjectResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:TagProject', tagProjectResponse_tags - The tags for the project.

$sel:httpStatus:TagProjectResponse', tagProjectResponse_httpStatus - The response's http status code.

DescribeProject

data DescribeProject

See: newDescribeProject smart constructor.

Constructors

DescribeProject' Text

newDescribeProject

Arguments

:: Text
$sel:id:DescribeProject'

-> DescribeProject

Create a value of DescribeProject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:DescribeProject', describeProject_id - The ID of the project.

data DescribeProjectResponse

See: newDescribeProjectResponse smart constructor.

Constructors

DescribeProjectResponse' (Maybe ProjectStatus) (Maybe Text) (Maybe Text) (Maybe (Sensitive Text)) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe POSIX) (Maybe (Sensitive Text)) Int

newDescribeProjectResponse

Arguments

:: Int
$sel:httpStatus:DescribeProjectResponse'

-> DescribeProjectResponse

Create a value of DescribeProjectResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DescribeProjectResponse', describeProjectResponse_status - The project creation or deletion status.

$sel:arn:DescribeProjectResponse', describeProjectResponse_arn - The Amazon Resource Name (ARN) for the project.

$sel:projectTemplateId:DescribeProjectResponse', describeProjectResponse_projectTemplateId - The ID for the AWS CodeStar project template used to create the project.

$sel:name:DescribeProjectResponse', describeProjectResponse_name - The display name for the project.

$sel:id:DescribeProject', describeProjectResponse_id - The ID of the project.

$sel:stackId:DescribeProjectResponse', describeProjectResponse_stackId - The ID of the primary stack in AWS CloudFormation used to generate resources for the project.

$sel:clientRequestToken:DescribeProjectResponse', describeProjectResponse_clientRequestToken - A user- or system-generated token that identifies the entity that requested project creation.

$sel:createdTimeStamp:DescribeProjectResponse', describeProjectResponse_createdTimeStamp - The date and time the project was created, in timestamp format.

$sel:description:DescribeProjectResponse', describeProjectResponse_description - The description of the project, if any.

$sel:httpStatus:DescribeProjectResponse', describeProjectResponse_httpStatus - The response's http status code.

ListUserProfiles (Paginated)

data ListUserProfiles

See: newListUserProfiles smart constructor.

Constructors

ListUserProfiles' (Maybe Text) (Maybe Natural)

newListUserProfiles :: ListUserProfiles

Create a value of ListUserProfiles with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListUserProfiles', listUserProfiles_nextToken - The continuation token for the next set of results, if the results cannot be returned in one response.

$sel:maxResults:ListUserProfiles', listUserProfiles_maxResults - The maximum number of results to return in a response.

data ListUserProfilesResponse

See: newListUserProfilesResponse smart constructor.

Constructors

ListUserProfilesResponse' (Maybe Text) Int [UserProfileSummary]

newListUserProfilesResponse

Arguments

:: Int
$sel:httpStatus:ListUserProfilesResponse'

-> ListUserProfilesResponse

Create a value of ListUserProfilesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListUserProfiles', listUserProfilesResponse_nextToken - The continuation token to use when requesting the next set of results, if there are more results to be returned.

$sel:httpStatus:ListUserProfilesResponse', listUserProfilesResponse_httpStatus - The response's http status code.

$sel:userProfiles:ListUserProfilesResponse', listUserProfilesResponse_userProfiles - All the user profiles configured in AWS CodeStar for an AWS account.

ListResources (Paginated)

data ListResources

See: newListResources smart constructor.

Constructors

ListResources' (Maybe Text) (Maybe Natural) Text

newListResources

Arguments

:: Text
$sel:projectId:ListResources'

-> ListResources

Create a value of ListResources with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListResources', listResources_nextToken - The continuation token for the next set of results, if the results cannot be returned in one response.

$sel:maxResults:ListResources', listResources_maxResults - The maximum amount of data that can be contained in a single set of results.

$sel:projectId:ListResources', listResources_projectId - The ID of the project.

data ListResourcesResponse

See: newListResourcesResponse smart constructor.

Constructors

ListResourcesResponse' (Maybe [Resource]) (Maybe Text) Int

newListResourcesResponse

Arguments

:: Int
$sel:httpStatus:ListResourcesResponse'

-> ListResourcesResponse

Create a value of ListResourcesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resources:ListResourcesResponse', listResourcesResponse_resources - An array of resources associated with the project.

$sel:nextToken:ListResources', listResourcesResponse_nextToken - The continuation token to use when requesting the next set of results, if there are more results to be returned.

$sel:httpStatus:ListResourcesResponse', listResourcesResponse_httpStatus - The response's http status code.

AssociateTeamMember

data AssociateTeamMember

See: newAssociateTeamMember smart constructor.

Constructors

AssociateTeamMember' (Maybe Bool) (Maybe Text) Text Text Text

newAssociateTeamMember

Arguments

:: Text
$sel:projectId:AssociateTeamMember'

-> Text
$sel:userArn:AssociateTeamMember'

-> Text
$sel:projectRole:AssociateTeamMember'

-> AssociateTeamMember

Create a value of AssociateTeamMember with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:remoteAccessAllowed:AssociateTeamMember', associateTeamMember_remoteAccessAllowed - Whether the team member is allowed to use an SSH public/private key pair to remotely access project resources, for example Amazon EC2 instances.

$sel:clientRequestToken:AssociateTeamMember', associateTeamMember_clientRequestToken - A user- or system-generated token that identifies the entity that requested the team member association to the project. This token can be used to repeat the request.

$sel:projectId:AssociateTeamMember', associateTeamMember_projectId - The ID of the project to which you will add the IAM user.

$sel:userArn:AssociateTeamMember', associateTeamMember_userArn - The Amazon Resource Name (ARN) for the IAM user you want to add to the AWS CodeStar project.

$sel:projectRole:AssociateTeamMember', associateTeamMember_projectRole - The AWS CodeStar project role that will apply to this user. This role determines what actions a user can take in an AWS CodeStar project.

data AssociateTeamMemberResponse

See: newAssociateTeamMemberResponse smart constructor.

Constructors

AssociateTeamMemberResponse' (Maybe Text) Int

newAssociateTeamMemberResponse

Arguments

:: Int
$sel:httpStatus:AssociateTeamMemberResponse'

-> AssociateTeamMemberResponse

Create a value of AssociateTeamMemberResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientRequestToken:AssociateTeamMember', associateTeamMemberResponse_clientRequestToken - The user- or system-generated token from the initial request that can be used to repeat the request.

$sel:httpStatus:AssociateTeamMemberResponse', associateTeamMemberResponse_httpStatus - The response's http status code.

UntagProject

data UntagProject

See: newUntagProject smart constructor.

Constructors

UntagProject' Text [Text]

newUntagProject

Arguments

:: Text
$sel:id:UntagProject'

-> UntagProject

Create a value of UntagProject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:UntagProject', untagProject_id - The ID of the project to remove tags from.

$sel:tags:UntagProject', untagProject_tags - The tags to remove from the project.

data UntagProjectResponse

See: newUntagProjectResponse smart constructor.

Constructors

UntagProjectResponse' Int

newUntagProjectResponse

Arguments

:: Int
$sel:httpStatus:UntagProjectResponse'

-> UntagProjectResponse

Create a value of UntagProjectResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UntagProjectResponse', untagProjectResponse_httpStatus - The response's http status code.

UpdateTeamMember

data UpdateTeamMember

See: newUpdateTeamMember smart constructor.

Constructors

UpdateTeamMember' (Maybe Bool) (Maybe Text) Text Text

newUpdateTeamMember

Arguments

:: Text
$sel:projectId:UpdateTeamMember'

-> Text
$sel:userArn:UpdateTeamMember'

-> UpdateTeamMember

Create a value of UpdateTeamMember with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:remoteAccessAllowed:UpdateTeamMember', updateTeamMember_remoteAccessAllowed - Whether a team member is allowed to remotely access project resources using the SSH public key associated with the user's profile. Even if this is set to True, the user must associate a public key with their profile before the user can access resources.

$sel:projectRole:UpdateTeamMember', updateTeamMember_projectRole - The role assigned to the user in the project. Project roles have different levels of access. For more information, see Working with Teams in the AWS CodeStar User Guide.

$sel:projectId:UpdateTeamMember', updateTeamMember_projectId - The ID of the project.

$sel:userArn:UpdateTeamMember', updateTeamMember_userArn - The Amazon Resource Name (ARN) of the user for whom you want to change team membership attributes.

data UpdateTeamMemberResponse

See: newUpdateTeamMemberResponse smart constructor.

Constructors

UpdateTeamMemberResponse' (Maybe Text) (Maybe Bool) (Maybe Text) Int

newUpdateTeamMemberResponse

Arguments

:: Int
$sel:httpStatus:UpdateTeamMemberResponse'

-> UpdateTeamMemberResponse

Create a value of UpdateTeamMemberResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userArn:UpdateTeamMember', updateTeamMemberResponse_userArn - The Amazon Resource Name (ARN) of the user whose team membership attributes were updated.

$sel:remoteAccessAllowed:UpdateTeamMember', updateTeamMemberResponse_remoteAccessAllowed - Whether a team member is allowed to remotely access project resources using the SSH public key associated with the user's profile.

$sel:projectRole:UpdateTeamMember', updateTeamMemberResponse_projectRole - The project role granted to the user.

$sel:httpStatus:UpdateTeamMemberResponse', updateTeamMemberResponse_httpStatus - The response's http status code.

DescribeUserProfile

data DescribeUserProfile

See: newDescribeUserProfile smart constructor.

Constructors

DescribeUserProfile' Text

newDescribeUserProfile

Arguments

:: Text
$sel:userArn:DescribeUserProfile'

-> DescribeUserProfile

Create a value of DescribeUserProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userArn:DescribeUserProfile', describeUserProfile_userArn - The Amazon Resource Name (ARN) of the user.

data DescribeUserProfileResponse

See: newDescribeUserProfileResponse smart constructor.

Constructors

DescribeUserProfileResponse' (Maybe Text) (Maybe (Sensitive Text)) (Maybe (Sensitive Text)) Int Text POSIX POSIX

newDescribeUserProfileResponse

Arguments

:: Int
$sel:httpStatus:DescribeUserProfileResponse'

-> Text
$sel:userArn:DescribeUserProfile'

-> UTCTime
$sel:createdTimestamp:DescribeUserProfileResponse'

-> UTCTime
$sel:lastModifiedTimestamp:DescribeUserProfileResponse'

-> DescribeUserProfileResponse

Create a value of DescribeUserProfileResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sshPublicKey:DescribeUserProfileResponse', describeUserProfileResponse_sshPublicKey - The SSH public key associated with the user. This SSH public key is associated with the user profile, and can be used in conjunction with the associated private key for access to project resources, such as Amazon EC2 instances, if a project owner grants remote access to those resources.

$sel:emailAddress:DescribeUserProfileResponse', describeUserProfileResponse_emailAddress - The email address for the user. Optional.

$sel:displayName:DescribeUserProfileResponse', describeUserProfileResponse_displayName - The display name shown for the user in AWS CodeStar projects. For example, this could be set to both first and last name ("Mary Major") or a single name ("Mary"). The display name is also used to generate the initial icon associated with the user in AWS CodeStar projects. If spaces are included in the display name, the first character that appears after the space will be used as the second character in the user initial icon. The initial icon displays a maximum of two characters, so a display name with more than one space (for example "Mary Jane Major") would generate an initial icon using the first character and the first character after the space ("MJ", not "MM").

$sel:httpStatus:DescribeUserProfileResponse', describeUserProfileResponse_httpStatus - The response's http status code.

$sel:userArn:DescribeUserProfile', describeUserProfileResponse_userArn - The Amazon Resource Name (ARN) of the user.

$sel:createdTimestamp:DescribeUserProfileResponse', describeUserProfileResponse_createdTimestamp - The date and time when the user profile was created in AWS CodeStar, in timestamp format.

$sel:lastModifiedTimestamp:DescribeUserProfileResponse', describeUserProfileResponse_lastModifiedTimestamp - The date and time when the user profile was last modified, in timestamp format.

ListTagsForProject

data ListTagsForProject

See: newListTagsForProject smart constructor.

Constructors

ListTagsForProject' (Maybe Text) (Maybe Natural) Text

newListTagsForProject

Arguments

:: Text
$sel:id:ListTagsForProject'

-> ListTagsForProject

Create a value of ListTagsForProject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListTagsForProject', listTagsForProject_nextToken - Reserved for future use.

$sel:maxResults:ListTagsForProject', listTagsForProject_maxResults - Reserved for future use.

$sel:id:ListTagsForProject', listTagsForProject_id - The ID of the project to get tags for.

data ListTagsForProjectResponse

See: newListTagsForProjectResponse smart constructor.

Constructors

ListTagsForProjectResponse' (Maybe Text) (Maybe (HashMap Text Text)) Int

newListTagsForProjectResponse

Arguments

:: Int
$sel:httpStatus:ListTagsForProjectResponse'

-> ListTagsForProjectResponse

Create a value of ListTagsForProjectResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListTagsForProject', listTagsForProjectResponse_nextToken - Reserved for future use.

$sel:tags:ListTagsForProjectResponse', listTagsForProjectResponse_tags - The tags for the project.

$sel:httpStatus:ListTagsForProjectResponse', listTagsForProjectResponse_httpStatus - The response's http status code.

DeleteUserProfile

data DeleteUserProfile

See: newDeleteUserProfile smart constructor.

Constructors

DeleteUserProfile' Text

newDeleteUserProfile

Arguments

:: Text
$sel:userArn:DeleteUserProfile'

-> DeleteUserProfile

Create a value of DeleteUserProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userArn:DeleteUserProfile', deleteUserProfile_userArn - The Amazon Resource Name (ARN) of the user to delete from AWS CodeStar.

data DeleteUserProfileResponse

See: newDeleteUserProfileResponse smart constructor.

Constructors

DeleteUserProfileResponse' Int Text

newDeleteUserProfileResponse

Arguments

:: Int
$sel:httpStatus:DeleteUserProfileResponse'

-> Text
$sel:userArn:DeleteUserProfile'

-> DeleteUserProfileResponse

Create a value of DeleteUserProfileResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteUserProfileResponse', deleteUserProfileResponse_httpStatus - The response's http status code.

$sel:userArn:DeleteUserProfile', deleteUserProfileResponse_userArn - The Amazon Resource Name (ARN) of the user deleted from AWS CodeStar.

UpdateUserProfile

data UpdateUserProfile

See: newUpdateUserProfile smart constructor.

Constructors

UpdateUserProfile' (Maybe Text) (Maybe (Sensitive Text)) (Maybe (Sensitive Text)) Text

newUpdateUserProfile

Arguments

:: Text
$sel:userArn:UpdateUserProfile'

-> UpdateUserProfile

Create a value of UpdateUserProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sshPublicKey:UpdateUserProfile', updateUserProfile_sshPublicKey - The SSH public key associated with the user in AWS CodeStar. If a project owner allows the user remote access to project resources, this public key will be used along with the user's private key for SSH access.

$sel:emailAddress:UpdateUserProfile', updateUserProfile_emailAddress - The email address that is displayed as part of the user's profile in AWS CodeStar.

$sel:displayName:UpdateUserProfile', updateUserProfile_displayName - The name that is displayed as the friendly name for the user in AWS CodeStar.

$sel:userArn:UpdateUserProfile', updateUserProfile_userArn - The name that will be displayed as the friendly name for the user in AWS CodeStar.

data UpdateUserProfileResponse

See: newUpdateUserProfileResponse smart constructor.

Constructors

UpdateUserProfileResponse' (Maybe POSIX) (Maybe Text) (Maybe (Sensitive Text)) (Maybe (Sensitive Text)) (Maybe POSIX) Int Text

newUpdateUserProfileResponse

Arguments

:: Int
$sel:httpStatus:UpdateUserProfileResponse'

-> Text
$sel:userArn:UpdateUserProfile'

-> UpdateUserProfileResponse

Create a value of UpdateUserProfileResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedTimestamp:UpdateUserProfileResponse', updateUserProfileResponse_lastModifiedTimestamp - The date the user profile was last modified, in timestamp format.

$sel:sshPublicKey:UpdateUserProfile', updateUserProfileResponse_sshPublicKey - The SSH public key associated with the user in AWS CodeStar. This is the public portion of the public/private keypair the user can use to access project resources if a project owner allows the user remote access to those resources.

$sel:emailAddress:UpdateUserProfile', updateUserProfileResponse_emailAddress - The email address that is displayed as part of the user's profile in AWS CodeStar.

$sel:displayName:UpdateUserProfile', updateUserProfileResponse_displayName - The name that is displayed as the friendly name for the user in AWS CodeStar.

$sel:createdTimestamp:UpdateUserProfileResponse', updateUserProfileResponse_createdTimestamp - The date the user profile was created, in timestamp format.

$sel:httpStatus:UpdateUserProfileResponse', updateUserProfileResponse_httpStatus - The response's http status code.

$sel:userArn:UpdateUserProfile', updateUserProfileResponse_userArn - The Amazon Resource Name (ARN) of the user in IAM.

CreateUserProfile

data CreateUserProfile

See: newCreateUserProfile smart constructor.

Constructors

CreateUserProfile' (Maybe Text) Text (Sensitive Text) (Sensitive Text)

newCreateUserProfile

Arguments

:: Text
$sel:userArn:CreateUserProfile'

-> Text
$sel:displayName:CreateUserProfile'

-> Text
$sel:emailAddress:CreateUserProfile'

-> CreateUserProfile

Create a value of CreateUserProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sshPublicKey:CreateUserProfile', createUserProfile_sshPublicKey - The SSH public key associated with the user in AWS CodeStar. If a project owner allows the user remote access to project resources, this public key will be used along with the user's private key for SSH access.

$sel:userArn:CreateUserProfile', createUserProfile_userArn - The Amazon Resource Name (ARN) of the user in IAM.

$sel:displayName:CreateUserProfile', createUserProfile_displayName - The name that will be displayed as the friendly name for the user in AWS CodeStar.

$sel:emailAddress:CreateUserProfile', createUserProfile_emailAddress - The email address that will be displayed as part of the user's profile in AWS CodeStar.

data CreateUserProfileResponse

See: newCreateUserProfileResponse smart constructor.

Constructors

CreateUserProfileResponse' (Maybe POSIX) (Maybe Text) (Maybe (Sensitive Text)) (Maybe (Sensitive Text)) (Maybe POSIX) Int Text

newCreateUserProfileResponse

Arguments

:: Int
$sel:httpStatus:CreateUserProfileResponse'

-> Text
$sel:userArn:CreateUserProfile'

-> CreateUserProfileResponse

Create a value of CreateUserProfileResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedTimestamp:CreateUserProfileResponse', createUserProfileResponse_lastModifiedTimestamp - The date the user profile was last modified, in timestamp format.

$sel:sshPublicKey:CreateUserProfile', createUserProfileResponse_sshPublicKey - The SSH public key associated with the user in AWS CodeStar. This is the public portion of the public/private keypair the user can use to access project resources if a project owner allows the user remote access to those resources.

$sel:emailAddress:CreateUserProfile', createUserProfileResponse_emailAddress - The email address that is displayed as part of the user's profile in AWS CodeStar.

$sel:displayName:CreateUserProfile', createUserProfileResponse_displayName - The name that is displayed as the friendly name for the user in AWS CodeStar.

$sel:createdTimestamp:CreateUserProfileResponse', createUserProfileResponse_createdTimestamp - The date the user profile was created, in timestamp format.

$sel:httpStatus:CreateUserProfileResponse', createUserProfileResponse_httpStatus - The response's http status code.

$sel:userArn:CreateUserProfile', createUserProfileResponse_userArn - The Amazon Resource Name (ARN) of the user in IAM.

CreateProject

data CreateProject

See: newCreateProject smart constructor.

Constructors

CreateProject' (Maybe [Code]) (Maybe Toolchain) (Maybe Text) (Maybe (Sensitive Text)) (Maybe (HashMap Text Text)) (Sensitive Text) Text

newCreateProject

Arguments

:: Text
$sel:name:CreateProject'

-> Text
$sel:id:CreateProject'

-> CreateProject

Create a value of CreateProject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceCode:CreateProject', createProject_sourceCode - A list of the Code objects submitted with the project request. If this parameter is specified, the request must also include the toolchain parameter.

$sel:toolchain:CreateProject', createProject_toolchain - The name of the toolchain template file submitted with the project request. If this parameter is specified, the request must also include the sourceCode parameter.

$sel:clientRequestToken:CreateProject', createProject_clientRequestToken - A user- or system-generated token that identifies the entity that requested project creation. This token can be used to repeat the request.

$sel:description:CreateProject', createProject_description - The description of the project, if any.

$sel:tags:CreateProject', createProject_tags - The tags created for the project.

$sel:name:CreateProject', createProject_name - The display name for the project to be created in AWS CodeStar.

$sel:id:CreateProject', createProject_id - The ID of the project to be created in AWS CodeStar.

data CreateProjectResponse

See: newCreateProjectResponse smart constructor.

Constructors

CreateProjectResponse' (Maybe Text) (Maybe Text) Int Text Text

newCreateProjectResponse

Arguments

:: Int
$sel:httpStatus:CreateProjectResponse'

-> Text
$sel:id:CreateProject'

-> Text
$sel:arn:CreateProjectResponse'

-> CreateProjectResponse

Create a value of CreateProjectResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:projectTemplateId:CreateProjectResponse', createProjectResponse_projectTemplateId - Reserved for future use.

$sel:clientRequestToken:CreateProject', createProjectResponse_clientRequestToken - A user- or system-generated token that identifies the entity that requested project creation.

$sel:httpStatus:CreateProjectResponse', createProjectResponse_httpStatus - The response's http status code.

$sel:id:CreateProject', createProjectResponse_id - The ID of the project.

$sel:arn:CreateProjectResponse', createProjectResponse_arn - The Amazon Resource Name (ARN) of the created project.

Types

Code

data Code

Location and destination information about the source code files provided with the project request. The source code is uploaded to the new project source repository after project creation.

See: newCode smart constructor.

Constructors

Code' CodeSource CodeDestination

newCode

Arguments

:: CodeSource
$sel:source:Code'

-> CodeDestination
$sel:destination:Code'

-> Code

Create a value of Code with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:source:Code', code_source - The location where the source code files provided with the project request are stored. AWS CodeStar retrieves the files during project creation.

$sel:destination:Code', code_destination - The repository to be created in AWS CodeStar. Valid values are AWS CodeCommit or GitHub. After AWS CodeStar provisions the new repository, the source code files provided with the project request are placed in the repository.

CodeCommitCodeDestination

data CodeCommitCodeDestination

Information about the AWS CodeCommit repository to be created in AWS CodeStar. This is where the source code files provided with the project request will be uploaded after project creation.

See: newCodeCommitCodeDestination smart constructor.

Constructors

CodeCommitCodeDestination' Text

newCodeCommitCodeDestination

Arguments

:: Text
$sel:name:CodeCommitCodeDestination'

-> CodeCommitCodeDestination

Create a value of CodeCommitCodeDestination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:CodeCommitCodeDestination', codeCommitCodeDestination_name - The name of the AWS CodeCommit repository to be created in AWS CodeStar.

CodeDestination

data CodeDestination

The repository to be created in AWS CodeStar. Valid values are AWS CodeCommit or GitHub. After AWS CodeStar provisions the new repository, the source code files provided with the project request are placed in the repository.

See: newCodeDestination smart constructor.

Constructors

CodeDestination' (Maybe CodeCommitCodeDestination) (Maybe GitHubCodeDestination)

newCodeDestination :: CodeDestination

Create a value of CodeDestination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:codeCommit:CodeDestination', codeDestination_codeCommit - Information about the AWS CodeCommit repository to be created in AWS CodeStar. This is where the source code files provided with the project request will be uploaded after project creation.

$sel:gitHub:CodeDestination', codeDestination_gitHub - Information about the GitHub repository to be created in AWS CodeStar. This is where the source code files provided with the project request will be uploaded after project creation.

CodeSource

data CodeSource

The location where the source code files provided with the project request are stored. AWS CodeStar retrieves the files during project creation.

See: newCodeSource smart constructor.

Constructors

CodeSource' S3Location

newCodeSource

Arguments

:: S3Location
$sel:s3:CodeSource'

-> CodeSource

Create a value of CodeSource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3:CodeSource', codeSource_s3 - Information about the Amazon S3 location where the source code files provided with the project request are stored.

GitHubCodeDestination

data GitHubCodeDestination

Information about the GitHub repository to be created in AWS CodeStar. This is where the source code files provided with the project request will be uploaded after project creation.

See: newGitHubCodeDestination smart constructor.

Constructors

GitHubCodeDestination' (Maybe Text) Text Text Text Bool Bool (Sensitive Text)

newGitHubCodeDestination

Arguments

:: Text
$sel:name:GitHubCodeDestination'

-> Text
$sel:type':GitHubCodeDestination'

-> Text
$sel:owner:GitHubCodeDestination'

-> Bool
$sel:privateRepository:GitHubCodeDestination'

-> Bool
$sel:issuesEnabled:GitHubCodeDestination'

-> Text
$sel:token:GitHubCodeDestination'

-> GitHubCodeDestination

Create a value of GitHubCodeDestination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:GitHubCodeDestination', gitHubCodeDestination_description - Description for the GitHub repository to be created in AWS CodeStar. This description displays in GitHub after the repository is created.

$sel:name:GitHubCodeDestination', gitHubCodeDestination_name - Name of the GitHub repository to be created in AWS CodeStar.

$sel:type':GitHubCodeDestination', gitHubCodeDestination_type - The type of GitHub repository to be created in AWS CodeStar. Valid values are User or Organization.

$sel:owner:GitHubCodeDestination', gitHubCodeDestination_owner - The GitHub username for the owner of the GitHub repository to be created in AWS CodeStar. If this repository should be owned by a GitHub organization, provide its name.

$sel:privateRepository:GitHubCodeDestination', gitHubCodeDestination_privateRepository - Whether the GitHub repository is to be a private repository.

$sel:issuesEnabled:GitHubCodeDestination', gitHubCodeDestination_issuesEnabled - Whether to enable issues for the GitHub repository.

$sel:token:GitHubCodeDestination', gitHubCodeDestination_token - The GitHub user's personal access token for the GitHub repository.

ProjectStatus

data ProjectStatus

An indication of whether a project creation or deletion is failed or successful.

See: newProjectStatus smart constructor.

Constructors

ProjectStatus' (Maybe Text) Text

newProjectStatus

Arguments

:: Text
$sel:state:ProjectStatus'

-> ProjectStatus

Create a value of ProjectStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reason:ProjectStatus', projectStatus_reason - In the case of a project creation or deletion failure, a reason for the failure.

$sel:state:ProjectStatus', projectStatus_state - The phase of completion for a project creation or deletion.

ProjectSummary

data ProjectSummary

Information about the metadata for a project.

See: newProjectSummary smart constructor.

Constructors

ProjectSummary' (Maybe Text) (Maybe Text)

newProjectSummary :: ProjectSummary

Create a value of ProjectSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:projectArn:ProjectSummary', projectSummary_projectArn - The Amazon Resource Name (ARN) of the project.

$sel:projectId:ProjectSummary', projectSummary_projectId - The ID of the project.

Resource

data Resource

Information about a resource for a project.

See: newResource smart constructor.

Constructors

Resource' Text

newResource

Arguments

:: Text
$sel:id:Resource'

-> Resource

Create a value of Resource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:Resource', resource_id - The Amazon Resource Name (ARN) of the resource.

S3Location

data S3Location

The Amazon S3 location where the source code files provided with the project request are stored.

See: newS3Location smart constructor.

Constructors

S3Location' (Maybe Text) (Maybe Text)

newS3Location :: S3Location

Create a value of S3Location with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bucketKey:S3Location', s3Location_bucketKey - The Amazon S3 object key where the source code files provided with the project request are stored.

$sel:bucketName:S3Location', s3Location_bucketName - The Amazon S3 bucket name where the source code files provided with the project request are stored.

TeamMember

data TeamMember

Information about a team member in a project.

See: newTeamMember smart constructor.

Constructors

TeamMember' (Maybe Bool) Text Text

newTeamMember

Arguments

:: Text
$sel:userArn:TeamMember'

-> Text
$sel:projectRole:TeamMember'

-> TeamMember

Create a value of TeamMember with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:remoteAccessAllowed:TeamMember', teamMember_remoteAccessAllowed - Whether the user is allowed to remotely access project resources using an SSH public/private key pair.

$sel:userArn:TeamMember', teamMember_userArn - The Amazon Resource Name (ARN) of the user in IAM.

$sel:projectRole:TeamMember', teamMember_projectRole - The role assigned to the user in the project. Project roles have different levels of access. For more information, see Working with Teams in the AWS CodeStar User Guide.

Toolchain

data Toolchain

The toolchain template file provided with the project request. AWS CodeStar uses the template to provision the toolchain stack in AWS CloudFormation.

See: newToolchain smart constructor.

Constructors

Toolchain' (Maybe (HashMap Text (Sensitive Text))) (Maybe Text) ToolchainSource

newToolchain

Arguments

:: ToolchainSource
$sel:source:Toolchain'

-> Toolchain

Create a value of Toolchain with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stackParameters:Toolchain', toolchain_stackParameters - The list of parameter overrides to be passed into the toolchain template during stack provisioning, if any.

$sel:roleArn:Toolchain', toolchain_roleArn - The service role ARN for AWS CodeStar to use for the toolchain template during stack provisioning.

$sel:source:Toolchain', toolchain_source - The Amazon S3 location where the toolchain template file provided with the project request is stored. AWS CodeStar retrieves the file during project creation.

ToolchainSource

data ToolchainSource

The Amazon S3 location where the toolchain template file provided with the project request is stored. AWS CodeStar retrieves the file during project creation.

See: newToolchainSource smart constructor.

Constructors

ToolchainSource' S3Location

newToolchainSource

Arguments

:: S3Location
$sel:s3:ToolchainSource'

-> ToolchainSource

Create a value of ToolchainSource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3:ToolchainSource', toolchainSource_s3 - The Amazon S3 bucket where the toolchain template file provided with the project request is stored.

UserProfileSummary

data UserProfileSummary

Information about a user's profile in AWS CodeStar.

See: newUserProfileSummary smart constructor.

Constructors

UserProfileSummary' (Maybe Text) (Maybe Text) (Maybe (Sensitive Text)) (Maybe (Sensitive Text))

newUserProfileSummary :: UserProfileSummary

Create a value of UserProfileSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sshPublicKey:UserProfileSummary', userProfileSummary_sshPublicKey - The SSH public key associated with the user in AWS CodeStar. If a project owner allows the user remote access to project resources, this public key will be used along with the user's private key for SSH access.

$sel:userArn:UserProfileSummary', userProfileSummary_userArn - The Amazon Resource Name (ARN) of the user in IAM.

$sel:emailAddress:UserProfileSummary', userProfileSummary_emailAddress - The email address associated with the user.

$sel:displayName:UserProfileSummary', userProfileSummary_displayName - The display name of a user in AWS CodeStar. For example, this could be set to both first and last name ("Mary Major") or a single name ("Mary"). The display name is also used to generate the initial icon associated with the user in AWS CodeStar projects. If spaces are included in the display name, the first character that appears after the space will be used as the second character in the user initial icon. The initial icon displays a maximum of two characters, so a display name with more than one space (for example "Mary Jane Major") would generate an initial icon using the first character and the first character after the space ("MJ", not "MM").