Network.AWS.CodeStar.AssociateTeamMember

Description

Adds an IAM user to the team for an AWS CodeStar project.

Creating a Request

data AssociateTeamMember

See: newAssociateTeamMember smart constructor.

Constructors

AssociateTeamMember'

Fields

remoteAccessAllowed :: Maybe Bool

Whether the team member is allowed to use an SSH public/private key pair to remotely access project resources, for example Amazon EC2 instances.

clientRequestToken :: Maybe Text

A user- or system-generated token that identifies the entity that requested the team member association to the project. This token can be used to repeat the request.

projectId :: Text

The ID of the project to which you will add the IAM user.

userArn :: Text

The Amazon Resource Name (ARN) for the IAM user you want to add to the AWS CodeStar project.

projectRole :: Text

The AWS CodeStar project role that will apply to this user. This role determines what actions a user can take in an AWS CodeStar project.

newAssociateTeamMember

Arguments

:: Text
$sel:projectId:AssociateTeamMember'

-> Text
$sel:userArn:AssociateTeamMember'

-> Text
$sel:projectRole:AssociateTeamMember'

-> AssociateTeamMember

Create a value of AssociateTeamMember with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:remoteAccessAllowed:AssociateTeamMember', associateTeamMember_remoteAccessAllowed - Whether the team member is allowed to use an SSH public/private key pair to remotely access project resources, for example Amazon EC2 instances.

$sel:clientRequestToken:AssociateTeamMember', associateTeamMember_clientRequestToken - A user- or system-generated token that identifies the entity that requested the team member association to the project. This token can be used to repeat the request.

$sel:projectId:AssociateTeamMember', associateTeamMember_projectId - The ID of the project to which you will add the IAM user.

$sel:userArn:AssociateTeamMember', associateTeamMember_userArn - The Amazon Resource Name (ARN) for the IAM user you want to add to the AWS CodeStar project.

$sel:projectRole:AssociateTeamMember', associateTeamMember_projectRole - The AWS CodeStar project role that will apply to this user. This role determines what actions a user can take in an AWS CodeStar project.

Request Lenses

associateTeamMember_remoteAccessAllowed :: Lens' AssociateTeamMember (Maybe Bool)

Whether the team member is allowed to use an SSH public/private key pair to remotely access project resources, for example Amazon EC2 instances.

associateTeamMember_clientRequestToken :: Lens' AssociateTeamMember (Maybe Text)

A user- or system-generated token that identifies the entity that requested the team member association to the project. This token can be used to repeat the request.

associateTeamMember_projectId :: Lens' AssociateTeamMember Text

The ID of the project to which you will add the IAM user.

associateTeamMember_userArn :: Lens' AssociateTeamMember Text

The Amazon Resource Name (ARN) for the IAM user you want to add to the AWS CodeStar project.

associateTeamMember_projectRole :: Lens' AssociateTeamMember Text

The AWS CodeStar project role that will apply to this user. This role determines what actions a user can take in an AWS CodeStar project.

Destructuring the Response

data AssociateTeamMemberResponse

See: newAssociateTeamMemberResponse smart constructor.

Constructors

AssociateTeamMemberResponse'

Fields

clientRequestToken :: Maybe Text

The user- or system-generated token from the initial request that can be used to repeat the request.

httpStatus :: Int

The response's http status code.

newAssociateTeamMemberResponse

Arguments

:: Int
$sel:httpStatus:AssociateTeamMemberResponse'

-> AssociateTeamMemberResponse

Create a value of AssociateTeamMemberResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientRequestToken:AssociateTeamMember', associateTeamMemberResponse_clientRequestToken - The user- or system-generated token from the initial request that can be used to repeat the request.

$sel:httpStatus:AssociateTeamMemberResponse', associateTeamMemberResponse_httpStatus - The response's http status code.

Response Lenses

associateTeamMemberResponse_clientRequestToken :: Lens' AssociateTeamMemberResponse (Maybe Text)

The user- or system-generated token from the initial request that can be used to repeat the request.

associateTeamMemberResponse_httpStatus :: Lens' AssociateTeamMemberResponse Int

The response's http status code.