Network.AWS.CodeStar.CreateProject

Description

Creates a project, including project resources. This action creates a project based on a submitted project request. A set of source code files and a toolchain template file can be included with the project request. If these are not provided, an empty project is created.

Creating a Request

data CreateProject

See: newCreateProject smart constructor.

Constructors

CreateProject'

Fields

sourceCode :: Maybe [Code]

A list of the Code objects submitted with the project request. If this parameter is specified, the request must also include the toolchain parameter.

toolchain :: Maybe Toolchain

The name of the toolchain template file submitted with the project request. If this parameter is specified, the request must also include the sourceCode parameter.

clientRequestToken :: Maybe Text

A user- or system-generated token that identifies the entity that requested project creation. This token can be used to repeat the request.

description :: Maybe (Sensitive Text)

The description of the project, if any.

tags :: Maybe (HashMap Text Text)

The tags created for the project.

name :: Sensitive Text

The display name for the project to be created in AWS CodeStar.

id :: Text

The ID of the project to be created in AWS CodeStar.

newCreateProject

Arguments

:: Text
$sel:name:CreateProject'

-> Text
$sel:id:CreateProject'

-> CreateProject

Create a value of CreateProject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceCode:CreateProject', createProject_sourceCode - A list of the Code objects submitted with the project request. If this parameter is specified, the request must also include the toolchain parameter.

$sel:toolchain:CreateProject', createProject_toolchain - The name of the toolchain template file submitted with the project request. If this parameter is specified, the request must also include the sourceCode parameter.

$sel:clientRequestToken:CreateProject', createProject_clientRequestToken - A user- or system-generated token that identifies the entity that requested project creation. This token can be used to repeat the request.

$sel:description:CreateProject', createProject_description - The description of the project, if any.

$sel:tags:CreateProject', createProject_tags - The tags created for the project.

$sel:name:CreateProject', createProject_name - The display name for the project to be created in AWS CodeStar.

$sel:id:CreateProject', createProject_id - The ID of the project to be created in AWS CodeStar.

Request Lenses

createProject_sourceCode :: Lens' CreateProject (Maybe [Code])

A list of the Code objects submitted with the project request. If this parameter is specified, the request must also include the toolchain parameter.

createProject_toolchain :: Lens' CreateProject (Maybe Toolchain)

The name of the toolchain template file submitted with the project request. If this parameter is specified, the request must also include the sourceCode parameter.

createProject_clientRequestToken :: Lens' CreateProject (Maybe Text)

A user- or system-generated token that identifies the entity that requested project creation. This token can be used to repeat the request.

createProject_description :: Lens' CreateProject (Maybe Text)

The description of the project, if any.

createProject_tags :: Lens' CreateProject (Maybe (HashMap Text Text))

The tags created for the project.

createProject_name :: Lens' CreateProject Text

The display name for the project to be created in AWS CodeStar.

createProject_id :: Lens' CreateProject Text

The ID of the project to be created in AWS CodeStar.

Destructuring the Response

data CreateProjectResponse

See: newCreateProjectResponse smart constructor.

Constructors

CreateProjectResponse'

Fields

projectTemplateId :: Maybe Text

Reserved for future use.

clientRequestToken :: Maybe Text

A user- or system-generated token that identifies the entity that requested project creation.

httpStatus :: Int

The response's http status code.

id :: Text

The ID of the project.

arn :: Text

The Amazon Resource Name (ARN) of the created project.

newCreateProjectResponse

Arguments

:: Int
$sel:httpStatus:CreateProjectResponse'

-> Text
$sel:id:CreateProject'

-> Text
$sel:arn:CreateProjectResponse'

-> CreateProjectResponse

Create a value of CreateProjectResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:projectTemplateId:CreateProjectResponse', createProjectResponse_projectTemplateId - Reserved for future use.

$sel:clientRequestToken:CreateProject', createProjectResponse_clientRequestToken - A user- or system-generated token that identifies the entity that requested project creation.

$sel:httpStatus:CreateProjectResponse', createProjectResponse_httpStatus - The response's http status code.

$sel:id:CreateProject', createProjectResponse_id - The ID of the project.

$sel:arn:CreateProjectResponse', createProjectResponse_arn - The Amazon Resource Name (ARN) of the created project.

Response Lenses

createProjectResponse_projectTemplateId :: Lens' CreateProjectResponse (Maybe Text)

Reserved for future use.

createProjectResponse_clientRequestToken :: Lens' CreateProjectResponse (Maybe Text)

A user- or system-generated token that identifies the entity that requested project creation.

createProjectResponse_httpStatus :: Lens' CreateProjectResponse Int

The response's http status code.

createProjectResponse_id :: Lens' CreateProjectResponse Text

The ID of the project.

createProjectResponse_arn :: Lens' CreateProjectResponse Text

The Amazon Resource Name (ARN) of the created project.