Network.AWS.CodeStar.DescribeUserProfile

Description

Describes a user in AWS CodeStar and the user attributes across all projects.

Creating a Request

data DescribeUserProfile

See: newDescribeUserProfile smart constructor.

Constructors

DescribeUserProfile'

Fields

userArn :: Text

The Amazon Resource Name (ARN) of the user.

newDescribeUserProfile

Arguments

:: Text
$sel:userArn:DescribeUserProfile'

-> DescribeUserProfile

Create a value of DescribeUserProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userArn:DescribeUserProfile', describeUserProfile_userArn - The Amazon Resource Name (ARN) of the user.

Request Lenses

describeUserProfile_userArn :: Lens' DescribeUserProfile Text

The Amazon Resource Name (ARN) of the user.

Destructuring the Response

data DescribeUserProfileResponse

See: newDescribeUserProfileResponse smart constructor.

Constructors

DescribeUserProfileResponse'

Fields

sshPublicKey :: Maybe Text

The SSH public key associated with the user. This SSH public key is associated with the user profile, and can be used in conjunction with the associated private key for access to project resources, such as Amazon EC2 instances, if a project owner grants remote access to those resources.

emailAddress :: Maybe (Sensitive Text)

The email address for the user. Optional.

displayName :: Maybe (Sensitive Text)

The display name shown for the user in AWS CodeStar projects. For example, this could be set to both first and last name ("Mary Major") or a single name ("Mary"). The display name is also used to generate the initial icon associated with the user in AWS CodeStar projects. If spaces are included in the display name, the first character that appears after the space will be used as the second character in the user initial icon. The initial icon displays a maximum of two characters, so a display name with more than one space (for example "Mary Jane Major") would generate an initial icon using the first character and the first character after the space ("MJ", not "MM").

httpStatus :: Int

The response's http status code.

userArn :: Text

The Amazon Resource Name (ARN) of the user.

createdTimestamp :: POSIX

The date and time when the user profile was created in AWS CodeStar, in timestamp format.

lastModifiedTimestamp :: POSIX

The date and time when the user profile was last modified, in timestamp format.

newDescribeUserProfileResponse

Arguments

:: Int
$sel:httpStatus:DescribeUserProfileResponse'

-> Text
$sel:userArn:DescribeUserProfile'

-> UTCTime
$sel:createdTimestamp:DescribeUserProfileResponse'

-> UTCTime
$sel:lastModifiedTimestamp:DescribeUserProfileResponse'

-> DescribeUserProfileResponse

Create a value of DescribeUserProfileResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sshPublicKey:DescribeUserProfileResponse', describeUserProfileResponse_sshPublicKey - The SSH public key associated with the user. This SSH public key is associated with the user profile, and can be used in conjunction with the associated private key for access to project resources, such as Amazon EC2 instances, if a project owner grants remote access to those resources.

$sel:emailAddress:DescribeUserProfileResponse', describeUserProfileResponse_emailAddress - The email address for the user. Optional.

$sel:displayName:DescribeUserProfileResponse', describeUserProfileResponse_displayName - The display name shown for the user in AWS CodeStar projects. For example, this could be set to both first and last name ("Mary Major") or a single name ("Mary"). The display name is also used to generate the initial icon associated with the user in AWS CodeStar projects. If spaces are included in the display name, the first character that appears after the space will be used as the second character in the user initial icon. The initial icon displays a maximum of two characters, so a display name with more than one space (for example "Mary Jane Major") would generate an initial icon using the first character and the first character after the space ("MJ", not "MM").

$sel:httpStatus:DescribeUserProfileResponse', describeUserProfileResponse_httpStatus - The response's http status code.

$sel:userArn:DescribeUserProfile', describeUserProfileResponse_userArn - The Amazon Resource Name (ARN) of the user.

$sel:createdTimestamp:DescribeUserProfileResponse', describeUserProfileResponse_createdTimestamp - The date and time when the user profile was created in AWS CodeStar, in timestamp format.

$sel:lastModifiedTimestamp:DescribeUserProfileResponse', describeUserProfileResponse_lastModifiedTimestamp - The date and time when the user profile was last modified, in timestamp format.

Response Lenses

describeUserProfileResponse_sshPublicKey :: Lens' DescribeUserProfileResponse (Maybe Text)

The SSH public key associated with the user. This SSH public key is associated with the user profile, and can be used in conjunction with the associated private key for access to project resources, such as Amazon EC2 instances, if a project owner grants remote access to those resources.

describeUserProfileResponse_emailAddress :: Lens' DescribeUserProfileResponse (Maybe Text)

The email address for the user. Optional.

describeUserProfileResponse_displayName :: Lens' DescribeUserProfileResponse (Maybe Text)

The display name shown for the user in AWS CodeStar projects. For example, this could be set to both first and last name ("Mary Major") or a single name ("Mary"). The display name is also used to generate the initial icon associated with the user in AWS CodeStar projects. If spaces are included in the display name, the first character that appears after the space will be used as the second character in the user initial icon. The initial icon displays a maximum of two characters, so a display name with more than one space (for example "Mary Jane Major") would generate an initial icon using the first character and the first character after the space ("MJ", not "MM").

describeUserProfileResponse_httpStatus :: Lens' DescribeUserProfileResponse Int

The response's http status code.

describeUserProfileResponse_userArn :: Lens' DescribeUserProfileResponse Text

The Amazon Resource Name (ARN) of the user.

describeUserProfileResponse_createdTimestamp :: Lens' DescribeUserProfileResponse UTCTime

The date and time when the user profile was created in AWS CodeStar, in timestamp format.

describeUserProfileResponse_lastModifiedTimestamp :: Lens' DescribeUserProfileResponse UTCTime

The date and time when the user profile was last modified, in timestamp format.