Network.AWS.CodeStar.DisassociateTeamMember

Description

Removes a user from a project. Removing a user from a project also removes the IAM policies from that user that allowed access to the project and its resources. Disassociating a team member does not remove that user's profile from AWS CodeStar. It does not remove the user from IAM.

Creating a Request

data DisassociateTeamMember

See: newDisassociateTeamMember smart constructor.

Constructors

DisassociateTeamMember'

Fields

projectId :: Text

The ID of the AWS CodeStar project from which you want to remove a team member.

userArn :: Text

The Amazon Resource Name (ARN) of the IAM user or group whom you want to remove from the project.

newDisassociateTeamMember

Arguments

:: Text
$sel:projectId:DisassociateTeamMember'

-> Text
$sel:userArn:DisassociateTeamMember'

-> DisassociateTeamMember

Create a value of DisassociateTeamMember with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:projectId:DisassociateTeamMember', disassociateTeamMember_projectId - The ID of the AWS CodeStar project from which you want to remove a team member.

$sel:userArn:DisassociateTeamMember', disassociateTeamMember_userArn - The Amazon Resource Name (ARN) of the IAM user or group whom you want to remove from the project.

Request Lenses

disassociateTeamMember_projectId :: Lens' DisassociateTeamMember Text

The ID of the AWS CodeStar project from which you want to remove a team member.

disassociateTeamMember_userArn :: Lens' DisassociateTeamMember Text

The Amazon Resource Name (ARN) of the IAM user or group whom you want to remove from the project.

Destructuring the Response

data DisassociateTeamMemberResponse

See: newDisassociateTeamMemberResponse smart constructor.

Constructors

DisassociateTeamMemberResponse'

Fields

httpStatus :: Int

The response's http status code.

newDisassociateTeamMemberResponse

Arguments

:: Int
$sel:httpStatus:DisassociateTeamMemberResponse'

-> DisassociateTeamMemberResponse

Create a value of DisassociateTeamMemberResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DisassociateTeamMemberResponse', disassociateTeamMemberResponse_httpStatus - The response's http status code.

Response Lenses

disassociateTeamMemberResponse_httpStatus :: Lens' DisassociateTeamMemberResponse Int

The response's http status code.