Network.AWS.CodeStar.Lens

Description

Operations

DescribeUserProfile

describeUserProfile_userArn :: Lens' DescribeUserProfile Text

The Amazon Resource Name (ARN) of the user.

describeUserProfileResponse_sshPublicKey :: Lens' DescribeUserProfileResponse (Maybe Text)

The SSH public key associated with the user. This SSH public key is associated with the user profile, and can be used in conjunction with the associated private key for access to project resources, such as Amazon EC2 instances, if a project owner grants remote access to those resources.

describeUserProfileResponse_displayName :: Lens' DescribeUserProfileResponse (Maybe Text)

The display name shown for the user in AWS CodeStar projects. For example, this could be set to both first and last name ("Mary Major") or a single name ("Mary"). The display name is also used to generate the initial icon associated with the user in AWS CodeStar projects. If spaces are included in the display name, the first character that appears after the space will be used as the second character in the user initial icon. The initial icon displays a maximum of two characters, so a display name with more than one space (for example "Mary Jane Major") would generate an initial icon using the first character and the first character after the space ("MJ", not "MM").

describeUserProfileResponse_emailAddress :: Lens' DescribeUserProfileResponse (Maybe Text)

The email address for the user. Optional.

describeUserProfileResponse_httpStatus :: Lens' DescribeUserProfileResponse Int

The response's http status code.

describeUserProfileResponse_userArn :: Lens' DescribeUserProfileResponse Text

The Amazon Resource Name (ARN) of the user.

describeUserProfileResponse_createdTimestamp :: Lens' DescribeUserProfileResponse UTCTime

The date and time when the user profile was created in AWS CodeStar, in timestamp format.

describeUserProfileResponse_lastModifiedTimestamp :: Lens' DescribeUserProfileResponse UTCTime

The date and time when the user profile was last modified, in timestamp format.

TagProject

tagProject_id :: Lens' TagProject Text

The ID of the project you want to add a tag to.

tagProject_tags :: Lens' TagProject (HashMap Text Text)

The tags you want to add to the project.

tagProjectResponse_tags :: Lens' TagProjectResponse (Maybe (HashMap Text Text))

The tags for the project.

tagProjectResponse_httpStatus :: Lens' TagProjectResponse Int

The response's http status code.

DisassociateTeamMember

disassociateTeamMember_projectId :: Lens' DisassociateTeamMember Text

The ID of the AWS CodeStar project from which you want to remove a team member.

disassociateTeamMember_userArn :: Lens' DisassociateTeamMember Text

The Amazon Resource Name (ARN) of the IAM user or group whom you want to remove from the project.

disassociateTeamMemberResponse_httpStatus :: Lens' DisassociateTeamMemberResponse Int

The response's http status code.

ListProjects

listProjects_nextToken :: Lens' ListProjects (Maybe Text)

The continuation token to be used to return the next set of results, if the results cannot be returned in one response.

listProjects_maxResults :: Lens' ListProjects (Maybe Natural)

The maximum amount of data that can be contained in a single set of results.

listProjectsResponse_nextToken :: Lens' ListProjectsResponse (Maybe Text)

The continuation token to use when requesting the next set of results, if there are more results to be returned.

listProjectsResponse_httpStatus :: Lens' ListProjectsResponse Int

The response's http status code.

listProjectsResponse_projects :: Lens' ListProjectsResponse [ProjectSummary]

A list of projects.

CreateProject

createProject_sourceCode :: Lens' CreateProject (Maybe [Code])

A list of the Code objects submitted with the project request. If this parameter is specified, the request must also include the toolchain parameter.

createProject_tags :: Lens' CreateProject (Maybe (HashMap Text Text))

The tags created for the project.

createProject_description :: Lens' CreateProject (Maybe Text)

The description of the project, if any.

createProject_clientRequestToken :: Lens' CreateProject (Maybe Text)

A user- or system-generated token that identifies the entity that requested project creation. This token can be used to repeat the request.

createProject_toolchain :: Lens' CreateProject (Maybe Toolchain)

The name of the toolchain template file submitted with the project request. If this parameter is specified, the request must also include the sourceCode parameter.

createProject_name :: Lens' CreateProject Text

The display name for the project to be created in AWS CodeStar.

createProject_id :: Lens' CreateProject Text

The ID of the project to be created in AWS CodeStar.

createProjectResponse_clientRequestToken :: Lens' CreateProjectResponse (Maybe Text)

A user- or system-generated token that identifies the entity that requested project creation.

createProjectResponse_projectTemplateId :: Lens' CreateProjectResponse (Maybe Text)

Reserved for future use.

createProjectResponse_httpStatus :: Lens' CreateProjectResponse Int

The response's http status code.

createProjectResponse_id :: Lens' CreateProjectResponse Text

The ID of the project.

createProjectResponse_arn :: Lens' CreateProjectResponse Text

The Amazon Resource Name (ARN) of the created project.

UpdateUserProfile

updateUserProfile_sshPublicKey :: Lens' UpdateUserProfile (Maybe Text)

The SSH public key associated with the user in AWS CodeStar. If a project owner allows the user remote access to project resources, this public key will be used along with the user's private key for SSH access.

updateUserProfile_displayName :: Lens' UpdateUserProfile (Maybe Text)

The name that is displayed as the friendly name for the user in AWS CodeStar.

updateUserProfile_emailAddress :: Lens' UpdateUserProfile (Maybe Text)

The email address that is displayed as part of the user's profile in AWS CodeStar.

updateUserProfile_userArn :: Lens' UpdateUserProfile Text

The name that will be displayed as the friendly name for the user in AWS CodeStar.

updateUserProfileResponse_createdTimestamp :: Lens' UpdateUserProfileResponse (Maybe UTCTime)

The date the user profile was created, in timestamp format.

updateUserProfileResponse_lastModifiedTimestamp :: Lens' UpdateUserProfileResponse (Maybe UTCTime)

The date the user profile was last modified, in timestamp format.

updateUserProfileResponse_sshPublicKey :: Lens' UpdateUserProfileResponse (Maybe Text)

The SSH public key associated with the user in AWS CodeStar. This is the public portion of the public/private keypair the user can use to access project resources if a project owner allows the user remote access to those resources.

updateUserProfileResponse_displayName :: Lens' UpdateUserProfileResponse (Maybe Text)

The name that is displayed as the friendly name for the user in AWS CodeStar.

updateUserProfileResponse_emailAddress :: Lens' UpdateUserProfileResponse (Maybe Text)

The email address that is displayed as part of the user's profile in AWS CodeStar.

updateUserProfileResponse_httpStatus :: Lens' UpdateUserProfileResponse Int

The response's http status code.

updateUserProfileResponse_userArn :: Lens' UpdateUserProfileResponse Text

The Amazon Resource Name (ARN) of the user in IAM.

ListUserProfiles

listUserProfiles_nextToken :: Lens' ListUserProfiles (Maybe Text)

The continuation token for the next set of results, if the results cannot be returned in one response.

listUserProfiles_maxResults :: Lens' ListUserProfiles (Maybe Natural)

The maximum number of results to return in a response.

listUserProfilesResponse_nextToken :: Lens' ListUserProfilesResponse (Maybe Text)

The continuation token to use when requesting the next set of results, if there are more results to be returned.

listUserProfilesResponse_httpStatus :: Lens' ListUserProfilesResponse Int

The response's http status code.

listUserProfilesResponse_userProfiles :: Lens' ListUserProfilesResponse [UserProfileSummary]

All the user profiles configured in AWS CodeStar for an AWS account.

DeleteUserProfile

deleteUserProfile_userArn :: Lens' DeleteUserProfile Text

The Amazon Resource Name (ARN) of the user to delete from AWS CodeStar.

deleteUserProfileResponse_httpStatus :: Lens' DeleteUserProfileResponse Int

The response's http status code.

deleteUserProfileResponse_userArn :: Lens' DeleteUserProfileResponse Text

The Amazon Resource Name (ARN) of the user deleted from AWS CodeStar.

ListTagsForProject

listTagsForProject_nextToken :: Lens' ListTagsForProject (Maybe Text)

Reserved for future use.

listTagsForProject_maxResults :: Lens' ListTagsForProject (Maybe Natural)

Reserved for future use.

listTagsForProject_id :: Lens' ListTagsForProject Text

The ID of the project to get tags for.

listTagsForProjectResponse_nextToken :: Lens' ListTagsForProjectResponse (Maybe Text)

Reserved for future use.

listTagsForProjectResponse_tags :: Lens' ListTagsForProjectResponse (Maybe (HashMap Text Text))

The tags for the project.

listTagsForProjectResponse_httpStatus :: Lens' ListTagsForProjectResponse Int

The response's http status code.

UpdateProject

updateProject_name :: Lens' UpdateProject (Maybe Text)

The name of the project you want to update.

updateProject_description :: Lens' UpdateProject (Maybe Text)

The description of the project, if any.

updateProject_id :: Lens' UpdateProject Text

The ID of the project you want to update.

updateProjectResponse_httpStatus :: Lens' UpdateProjectResponse Int

The response's http status code.

DeleteProject

deleteProject_deleteStack :: Lens' DeleteProject (Maybe Bool)

Whether to send a delete request for the primary stack in AWS CloudFormation originally used to generate the project and its resources. This option will delete all AWS resources for the project (except for any buckets in Amazon S3) as well as deleting the project itself. Recommended for most use cases.

deleteProject_clientRequestToken :: Lens' DeleteProject (Maybe Text)

A user- or system-generated token that identifies the entity that requested project deletion. This token can be used to repeat the request.

deleteProject_id :: Lens' DeleteProject Text

The ID of the project to be deleted in AWS CodeStar.

deleteProjectResponse_stackId :: Lens' DeleteProjectResponse (Maybe Text)

The ID of the primary stack in AWS CloudFormation that will be deleted as part of deleting the project and its resources.

deleteProjectResponse_projectArn :: Lens' DeleteProjectResponse (Maybe Text)

The Amazon Resource Name (ARN) of the deleted project.

deleteProjectResponse_httpStatus :: Lens' DeleteProjectResponse Int

The response's http status code.

UntagProject

untagProject_id :: Lens' UntagProject Text

The ID of the project to remove tags from.

untagProject_tags :: Lens' UntagProject [Text]

The tags to remove from the project.

untagProjectResponse_httpStatus :: Lens' UntagProjectResponse Int

The response's http status code.

ListTeamMembers

listTeamMembers_nextToken :: Lens' ListTeamMembers (Maybe Text)

The continuation token for the next set of results, if the results cannot be returned in one response.

listTeamMembers_maxResults :: Lens' ListTeamMembers (Maybe Natural)

The maximum number of team members you want returned in a response.

listTeamMembers_projectId :: Lens' ListTeamMembers Text

The ID of the project for which you want to list team members.

listTeamMembersResponse_nextToken :: Lens' ListTeamMembersResponse (Maybe Text)

The continuation token to use when requesting the next set of results, if there are more results to be returned.

listTeamMembersResponse_httpStatus :: Lens' ListTeamMembersResponse Int

The response's http status code.

listTeamMembersResponse_teamMembers :: Lens' ListTeamMembersResponse [TeamMember]

A list of team member objects for the project.

UpdateTeamMember

updateTeamMember_projectRole :: Lens' UpdateTeamMember (Maybe Text)

The role assigned to the user in the project. Project roles have different levels of access. For more information, see Working with Teams in the AWS CodeStar User Guide.

updateTeamMember_remoteAccessAllowed :: Lens' UpdateTeamMember (Maybe Bool)

Whether a team member is allowed to remotely access project resources using the SSH public key associated with the user's profile. Even if this is set to True, the user must associate a public key with their profile before the user can access resources.

updateTeamMember_projectId :: Lens' UpdateTeamMember Text

The ID of the project.

updateTeamMember_userArn :: Lens' UpdateTeamMember Text

The Amazon Resource Name (ARN) of the user for whom you want to change team membership attributes.

updateTeamMemberResponse_userArn :: Lens' UpdateTeamMemberResponse (Maybe Text)

The Amazon Resource Name (ARN) of the user whose team membership attributes were updated.

updateTeamMemberResponse_projectRole :: Lens' UpdateTeamMemberResponse (Maybe Text)

The project role granted to the user.

updateTeamMemberResponse_remoteAccessAllowed :: Lens' UpdateTeamMemberResponse (Maybe Bool)

Whether a team member is allowed to remotely access project resources using the SSH public key associated with the user's profile.

updateTeamMemberResponse_httpStatus :: Lens' UpdateTeamMemberResponse Int

The response's http status code.

AssociateTeamMember

associateTeamMember_remoteAccessAllowed :: Lens' AssociateTeamMember (Maybe Bool)

Whether the team member is allowed to use an SSH public/private key pair to remotely access project resources, for example Amazon EC2 instances.

associateTeamMember_clientRequestToken :: Lens' AssociateTeamMember (Maybe Text)

A user- or system-generated token that identifies the entity that requested the team member association to the project. This token can be used to repeat the request.

associateTeamMember_projectId :: Lens' AssociateTeamMember Text

The ID of the project to which you will add the IAM user.

associateTeamMember_userArn :: Lens' AssociateTeamMember Text

The Amazon Resource Name (ARN) for the IAM user you want to add to the AWS CodeStar project.

associateTeamMember_projectRole :: Lens' AssociateTeamMember Text

The AWS CodeStar project role that will apply to this user. This role determines what actions a user can take in an AWS CodeStar project.

associateTeamMemberResponse_clientRequestToken :: Lens' AssociateTeamMemberResponse (Maybe Text)

The user- or system-generated token from the initial request that can be used to repeat the request.

associateTeamMemberResponse_httpStatus :: Lens' AssociateTeamMemberResponse Int

The response's http status code.

CreateUserProfile

createUserProfile_sshPublicKey :: Lens' CreateUserProfile (Maybe Text)

The SSH public key associated with the user in AWS CodeStar. If a project owner allows the user remote access to project resources, this public key will be used along with the user's private key for SSH access.

createUserProfile_userArn :: Lens' CreateUserProfile Text

The Amazon Resource Name (ARN) of the user in IAM.

createUserProfile_displayName :: Lens' CreateUserProfile Text

The name that will be displayed as the friendly name for the user in AWS CodeStar.

createUserProfile_emailAddress :: Lens' CreateUserProfile Text

The email address that will be displayed as part of the user's profile in AWS CodeStar.

createUserProfileResponse_createdTimestamp :: Lens' CreateUserProfileResponse (Maybe UTCTime)

The date the user profile was created, in timestamp format.

createUserProfileResponse_lastModifiedTimestamp :: Lens' CreateUserProfileResponse (Maybe UTCTime)

The date the user profile was last modified, in timestamp format.

createUserProfileResponse_sshPublicKey :: Lens' CreateUserProfileResponse (Maybe Text)

The SSH public key associated with the user in AWS CodeStar. This is the public portion of the public/private keypair the user can use to access project resources if a project owner allows the user remote access to those resources.

createUserProfileResponse_displayName :: Lens' CreateUserProfileResponse (Maybe Text)

The name that is displayed as the friendly name for the user in AWS CodeStar.

createUserProfileResponse_emailAddress :: Lens' CreateUserProfileResponse (Maybe Text)

The email address that is displayed as part of the user's profile in AWS CodeStar.

createUserProfileResponse_httpStatus :: Lens' CreateUserProfileResponse Int

The response's http status code.

createUserProfileResponse_userArn :: Lens' CreateUserProfileResponse Text

The Amazon Resource Name (ARN) of the user in IAM.

ListResources

listResources_nextToken :: Lens' ListResources (Maybe Text)

The continuation token for the next set of results, if the results cannot be returned in one response.

listResources_maxResults :: Lens' ListResources (Maybe Natural)

The maximum amount of data that can be contained in a single set of results.

listResources_projectId :: Lens' ListResources Text

The ID of the project.

listResourcesResponse_nextToken :: Lens' ListResourcesResponse (Maybe Text)

The continuation token to use when requesting the next set of results, if there are more results to be returned.

listResourcesResponse_resources :: Lens' ListResourcesResponse (Maybe [Resource])

An array of resources associated with the project.

listResourcesResponse_httpStatus :: Lens' ListResourcesResponse Int

The response's http status code.

DescribeProject

describeProject_id :: Lens' DescribeProject Text

The ID of the project.

describeProjectResponse_status :: Lens' DescribeProjectResponse (Maybe ProjectStatus)

The project creation or deletion status.

describeProjectResponse_createdTimeStamp :: Lens' DescribeProjectResponse (Maybe UTCTime)

The date and time the project was created, in timestamp format.

describeProjectResponse_stackId :: Lens' DescribeProjectResponse (Maybe Text)

The ID of the primary stack in AWS CloudFormation used to generate resources for the project.

describeProjectResponse_id :: Lens' DescribeProjectResponse (Maybe Text)

The ID of the project.

describeProjectResponse_arn :: Lens' DescribeProjectResponse (Maybe Text)

The Amazon Resource Name (ARN) for the project.

describeProjectResponse_name :: Lens' DescribeProjectResponse (Maybe Text)

The display name for the project.

describeProjectResponse_description :: Lens' DescribeProjectResponse (Maybe Text)

The description of the project, if any.

describeProjectResponse_clientRequestToken :: Lens' DescribeProjectResponse (Maybe Text)

A user- or system-generated token that identifies the entity that requested project creation.

describeProjectResponse_projectTemplateId :: Lens' DescribeProjectResponse (Maybe Text)

The ID for the AWS CodeStar project template used to create the project.

describeProjectResponse_httpStatus :: Lens' DescribeProjectResponse Int

The response's http status code.

Types

Code

code_source :: Lens' Code CodeSource

The location where the source code files provided with the project request are stored. AWS CodeStar retrieves the files during project creation.

code_destination :: Lens' Code CodeDestination

The repository to be created in AWS CodeStar. Valid values are AWS CodeCommit or GitHub. After AWS CodeStar provisions the new repository, the source code files provided with the project request are placed in the repository.

CodeCommitCodeDestination

codeCommitCodeDestination_name :: Lens' CodeCommitCodeDestination Text

The name of the AWS CodeCommit repository to be created in AWS CodeStar.

CodeDestination

codeDestination_codeCommit :: Lens' CodeDestination (Maybe CodeCommitCodeDestination)

Information about the AWS CodeCommit repository to be created in AWS CodeStar. This is where the source code files provided with the project request will be uploaded after project creation.

codeDestination_gitHub :: Lens' CodeDestination (Maybe GitHubCodeDestination)

Information about the GitHub repository to be created in AWS CodeStar. This is where the source code files provided with the project request will be uploaded after project creation.

CodeSource

codeSource_s3 :: Lens' CodeSource S3Location

Information about the Amazon S3 location where the source code files provided with the project request are stored.

GitHubCodeDestination

gitHubCodeDestination_description :: Lens' GitHubCodeDestination (Maybe Text)

Description for the GitHub repository to be created in AWS CodeStar. This description displays in GitHub after the repository is created.

gitHubCodeDestination_name :: Lens' GitHubCodeDestination Text

Name of the GitHub repository to be created in AWS CodeStar.

gitHubCodeDestination_type :: Lens' GitHubCodeDestination Text

The type of GitHub repository to be created in AWS CodeStar. Valid values are User or Organization.

gitHubCodeDestination_owner :: Lens' GitHubCodeDestination Text

The GitHub username for the owner of the GitHub repository to be created in AWS CodeStar. If this repository should be owned by a GitHub organization, provide its name.

gitHubCodeDestination_privateRepository :: Lens' GitHubCodeDestination Bool

Whether the GitHub repository is to be a private repository.

gitHubCodeDestination_issuesEnabled :: Lens' GitHubCodeDestination Bool

Whether to enable issues for the GitHub repository.

gitHubCodeDestination_token :: Lens' GitHubCodeDestination Text

The GitHub user's personal access token for the GitHub repository.

ProjectStatus

projectStatus_reason :: Lens' ProjectStatus (Maybe Text)

In the case of a project creation or deletion failure, a reason for the failure.

projectStatus_state :: Lens' ProjectStatus Text

The phase of completion for a project creation or deletion.

ProjectSummary

projectSummary_projectId :: Lens' ProjectSummary (Maybe Text)

The ID of the project.

projectSummary_projectArn :: Lens' ProjectSummary (Maybe Text)

The Amazon Resource Name (ARN) of the project.

Resource

resource_id :: Lens' Resource Text

The Amazon Resource Name (ARN) of the resource.

S3Location

s3Location_bucketName :: Lens' S3Location (Maybe Text)

The Amazon S3 bucket name where the source code files provided with the project request are stored.

s3Location_bucketKey :: Lens' S3Location (Maybe Text)

The Amazon S3 object key where the source code files provided with the project request are stored.

TeamMember

teamMember_remoteAccessAllowed :: Lens' TeamMember (Maybe Bool)

Whether the user is allowed to remotely access project resources using an SSH public/private key pair.

teamMember_userArn :: Lens' TeamMember Text

The Amazon Resource Name (ARN) of the user in IAM.

teamMember_projectRole :: Lens' TeamMember Text

The role assigned to the user in the project. Project roles have different levels of access. For more information, see Working with Teams in the AWS CodeStar User Guide.

Toolchain

toolchain_stackParameters :: Lens' Toolchain (Maybe (HashMap Text Text))

The list of parameter overrides to be passed into the toolchain template during stack provisioning, if any.

toolchain_roleArn :: Lens' Toolchain (Maybe Text)

The service role ARN for AWS CodeStar to use for the toolchain template during stack provisioning.

toolchain_source :: Lens' Toolchain ToolchainSource

The Amazon S3 location where the toolchain template file provided with the project request is stored. AWS CodeStar retrieves the file during project creation.

ToolchainSource

toolchainSource_s3 :: Lens' ToolchainSource S3Location

The Amazon S3 bucket where the toolchain template file provided with the project request is stored.

UserProfileSummary

userProfileSummary_userArn :: Lens' UserProfileSummary (Maybe Text)

The Amazon Resource Name (ARN) of the user in IAM.

userProfileSummary_sshPublicKey :: Lens' UserProfileSummary (Maybe Text)

The SSH public key associated with the user in AWS CodeStar. If a project owner allows the user remote access to project resources, this public key will be used along with the user's private key for SSH access.

userProfileSummary_displayName :: Lens' UserProfileSummary (Maybe Text)

The display name of a user in AWS CodeStar. For example, this could be set to both first and last name ("Mary Major") or a single name ("Mary"). The display name is also used to generate the initial icon associated with the user in AWS CodeStar projects. If spaces are included in the display name, the first character that appears after the space will be used as the second character in the user initial icon. The initial icon displays a maximum of two characters, so a display name with more than one space (for example "Mary Jane Major") would generate an initial icon using the first character and the first character after the space ("MJ", not "MM").

userProfileSummary_emailAddress :: Lens' UserProfileSummary (Maybe Text)

The email address associated with the user.