Network.AWS.CodeStar.Types

Description

Service Configuration

defaultService :: Service

API version 2017-04-19 of the Amazon CodeStar SDK configuration.

Errors

_TeamMemberAlreadyAssociatedException :: AsError a => Getting (First ServiceError) a ServiceError

The team member is already associated with a role in this project.

_ProjectAlreadyExistsException :: AsError a => Getting (First ServiceError) a ServiceError

An AWS CodeStar project with the same ID already exists in this region for the AWS account. AWS CodeStar project IDs must be unique within a region for the AWS account.

_ProjectNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified AWS CodeStar project was not found.

_UserProfileNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The user profile was not found.

_ProjectCreationFailedException :: AsError a => Getting (First ServiceError) a ServiceError

The project creation request was valid, but a nonspecific exception or error occurred during project creation. The project could not be created in AWS CodeStar.

_ConcurrentModificationException :: AsError a => Getting (First ServiceError) a ServiceError

Another modification is being made. That modification must complete before you can make your change.

_InvalidNextTokenException :: AsError a => Getting (First ServiceError) a ServiceError

The next token is not valid.

_ProjectConfigurationException :: AsError a => Getting (First ServiceError) a ServiceError

Project configuration information is required but not specified.

_InvalidServiceRoleException :: AsError a => Getting (First ServiceError) a ServiceError

The service role is not valid.

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The specified input is either not valid, or it could not be validated.

_TeamMemberNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified team member was not found.

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

A resource limit has been exceeded.

_UserProfileAlreadyExistsException :: AsError a => Getting (First ServiceError) a ServiceError

A user profile with that name already exists in this region for the AWS account. AWS CodeStar user profile names must be unique within a region for the AWS account.

Code

data Code

Location and destination information about the source code files provided with the project request. The source code is uploaded to the new project source repository after project creation.

See: newCode smart constructor.

Constructors

Code'

Fields

source :: CodeSource

The location where the source code files provided with the project request are stored. AWS CodeStar retrieves the files during project creation.

destination :: CodeDestination

The repository to be created in AWS CodeStar. Valid values are AWS CodeCommit or GitHub. After AWS CodeStar provisions the new repository, the source code files provided with the project request are placed in the repository.

newCode

Arguments

:: CodeSource
$sel:source:Code'

-> CodeDestination
$sel:destination:Code'

-> Code

Create a value of Code with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:source:Code', code_source - The location where the source code files provided with the project request are stored. AWS CodeStar retrieves the files during project creation.

$sel:destination:Code', code_destination - The repository to be created in AWS CodeStar. Valid values are AWS CodeCommit or GitHub. After AWS CodeStar provisions the new repository, the source code files provided with the project request are placed in the repository.

code_source :: Lens' Code CodeSource

The location where the source code files provided with the project request are stored. AWS CodeStar retrieves the files during project creation.

code_destination :: Lens' Code CodeDestination

The repository to be created in AWS CodeStar. Valid values are AWS CodeCommit or GitHub. After AWS CodeStar provisions the new repository, the source code files provided with the project request are placed in the repository.

CodeCommitCodeDestination

data CodeCommitCodeDestination

Information about the AWS CodeCommit repository to be created in AWS CodeStar. This is where the source code files provided with the project request will be uploaded after project creation.

See: newCodeCommitCodeDestination smart constructor.

Constructors

CodeCommitCodeDestination'

Fields

name :: Text

The name of the AWS CodeCommit repository to be created in AWS CodeStar.

newCodeCommitCodeDestination

Arguments

:: Text
$sel:name:CodeCommitCodeDestination'

-> CodeCommitCodeDestination

Create a value of CodeCommitCodeDestination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:CodeCommitCodeDestination', codeCommitCodeDestination_name - The name of the AWS CodeCommit repository to be created in AWS CodeStar.

codeCommitCodeDestination_name :: Lens' CodeCommitCodeDestination Text

The name of the AWS CodeCommit repository to be created in AWS CodeStar.

CodeDestination

data CodeDestination

The repository to be created in AWS CodeStar. Valid values are AWS CodeCommit or GitHub. After AWS CodeStar provisions the new repository, the source code files provided with the project request are placed in the repository.

See: newCodeDestination smart constructor.

Constructors

CodeDestination'

Fields

codeCommit :: Maybe CodeCommitCodeDestination

Information about the AWS CodeCommit repository to be created in AWS CodeStar. This is where the source code files provided with the project request will be uploaded after project creation.

gitHub :: Maybe GitHubCodeDestination

Information about the GitHub repository to be created in AWS CodeStar. This is where the source code files provided with the project request will be uploaded after project creation.

newCodeDestination :: CodeDestination

Create a value of CodeDestination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:codeCommit:CodeDestination', codeDestination_codeCommit - Information about the AWS CodeCommit repository to be created in AWS CodeStar. This is where the source code files provided with the project request will be uploaded after project creation.

$sel:gitHub:CodeDestination', codeDestination_gitHub - Information about the GitHub repository to be created in AWS CodeStar. This is where the source code files provided with the project request will be uploaded after project creation.

codeDestination_codeCommit :: Lens' CodeDestination (Maybe CodeCommitCodeDestination)

Information about the AWS CodeCommit repository to be created in AWS CodeStar. This is where the source code files provided with the project request will be uploaded after project creation.

codeDestination_gitHub :: Lens' CodeDestination (Maybe GitHubCodeDestination)

Information about the GitHub repository to be created in AWS CodeStar. This is where the source code files provided with the project request will be uploaded after project creation.

CodeSource

data CodeSource

The location where the source code files provided with the project request are stored. AWS CodeStar retrieves the files during project creation.

See: newCodeSource smart constructor.

Constructors

CodeSource'

Fields

s3 :: S3Location

Information about the Amazon S3 location where the source code files provided with the project request are stored.

newCodeSource

Arguments

:: S3Location
$sel:s3:CodeSource'

-> CodeSource

Create a value of CodeSource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3:CodeSource', codeSource_s3 - Information about the Amazon S3 location where the source code files provided with the project request are stored.

codeSource_s3 :: Lens' CodeSource S3Location

Information about the Amazon S3 location where the source code files provided with the project request are stored.

GitHubCodeDestination

data GitHubCodeDestination

Information about the GitHub repository to be created in AWS CodeStar. This is where the source code files provided with the project request will be uploaded after project creation.

See: newGitHubCodeDestination smart constructor.

Constructors

GitHubCodeDestination'

Fields

description :: Maybe Text

Description for the GitHub repository to be created in AWS CodeStar. This description displays in GitHub after the repository is created.

name :: Text

Name of the GitHub repository to be created in AWS CodeStar.

type' :: Text

The type of GitHub repository to be created in AWS CodeStar. Valid values are User or Organization.

owner :: Text

The GitHub username for the owner of the GitHub repository to be created in AWS CodeStar. If this repository should be owned by a GitHub organization, provide its name.

privateRepository :: Bool

Whether the GitHub repository is to be a private repository.

issuesEnabled :: Bool

Whether to enable issues for the GitHub repository.

token :: Sensitive Text

The GitHub user's personal access token for the GitHub repository.

newGitHubCodeDestination

Arguments

:: Text
$sel:name:GitHubCodeDestination'

-> Text
$sel:type':GitHubCodeDestination'

-> Text
$sel:owner:GitHubCodeDestination'

-> Bool
$sel:privateRepository:GitHubCodeDestination'

-> Bool
$sel:issuesEnabled:GitHubCodeDestination'

-> Text
$sel:token:GitHubCodeDestination'

-> GitHubCodeDestination

Create a value of GitHubCodeDestination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:GitHubCodeDestination', gitHubCodeDestination_description - Description for the GitHub repository to be created in AWS CodeStar. This description displays in GitHub after the repository is created.

$sel:name:GitHubCodeDestination', gitHubCodeDestination_name - Name of the GitHub repository to be created in AWS CodeStar.

$sel:type':GitHubCodeDestination', gitHubCodeDestination_type - The type of GitHub repository to be created in AWS CodeStar. Valid values are User or Organization.

$sel:owner:GitHubCodeDestination', gitHubCodeDestination_owner - The GitHub username for the owner of the GitHub repository to be created in AWS CodeStar. If this repository should be owned by a GitHub organization, provide its name.

$sel:privateRepository:GitHubCodeDestination', gitHubCodeDestination_privateRepository - Whether the GitHub repository is to be a private repository.

$sel:issuesEnabled:GitHubCodeDestination', gitHubCodeDestination_issuesEnabled - Whether to enable issues for the GitHub repository.

$sel:token:GitHubCodeDestination', gitHubCodeDestination_token - The GitHub user's personal access token for the GitHub repository.

gitHubCodeDestination_description :: Lens' GitHubCodeDestination (Maybe Text)

Description for the GitHub repository to be created in AWS CodeStar. This description displays in GitHub after the repository is created.

gitHubCodeDestination_name :: Lens' GitHubCodeDestination Text

Name of the GitHub repository to be created in AWS CodeStar.

gitHubCodeDestination_type :: Lens' GitHubCodeDestination Text

The type of GitHub repository to be created in AWS CodeStar. Valid values are User or Organization.

gitHubCodeDestination_owner :: Lens' GitHubCodeDestination Text

The GitHub username for the owner of the GitHub repository to be created in AWS CodeStar. If this repository should be owned by a GitHub organization, provide its name.

gitHubCodeDestination_privateRepository :: Lens' GitHubCodeDestination Bool

Whether the GitHub repository is to be a private repository.

gitHubCodeDestination_issuesEnabled :: Lens' GitHubCodeDestination Bool

Whether to enable issues for the GitHub repository.

gitHubCodeDestination_token :: Lens' GitHubCodeDestination Text

The GitHub user's personal access token for the GitHub repository.

ProjectStatus

data ProjectStatus

An indication of whether a project creation or deletion is failed or successful.

See: newProjectStatus smart constructor.

Constructors

ProjectStatus'

Fields

reason :: Maybe Text

In the case of a project creation or deletion failure, a reason for the failure.

state :: Text

The phase of completion for a project creation or deletion.

newProjectStatus

Arguments

:: Text
$sel:state:ProjectStatus'

-> ProjectStatus

Create a value of ProjectStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reason:ProjectStatus', projectStatus_reason - In the case of a project creation or deletion failure, a reason for the failure.

$sel:state:ProjectStatus', projectStatus_state - The phase of completion for a project creation or deletion.

projectStatus_reason :: Lens' ProjectStatus (Maybe Text)

In the case of a project creation or deletion failure, a reason for the failure.

projectStatus_state :: Lens' ProjectStatus Text

The phase of completion for a project creation or deletion.

ProjectSummary

data ProjectSummary

Information about the metadata for a project.

See: newProjectSummary smart constructor.

Constructors

ProjectSummary'

Fields

projectId :: Maybe Text

The ID of the project.

projectArn :: Maybe Text

The Amazon Resource Name (ARN) of the project.

newProjectSummary :: ProjectSummary

Create a value of ProjectSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:projectId:ProjectSummary', projectSummary_projectId - The ID of the project.

$sel:projectArn:ProjectSummary', projectSummary_projectArn - The Amazon Resource Name (ARN) of the project.

projectSummary_projectId :: Lens' ProjectSummary (Maybe Text)

The ID of the project.

projectSummary_projectArn :: Lens' ProjectSummary (Maybe Text)

The Amazon Resource Name (ARN) of the project.

Resource

data Resource

Information about a resource for a project.

See: newResource smart constructor.

Constructors

Resource'

Fields

id :: Text

The Amazon Resource Name (ARN) of the resource.

newResource

Arguments

:: Text
$sel:id:Resource'

-> Resource

Create a value of Resource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:Resource', resource_id - The Amazon Resource Name (ARN) of the resource.

resource_id :: Lens' Resource Text

The Amazon Resource Name (ARN) of the resource.

S3Location

data S3Location

The Amazon S3 location where the source code files provided with the project request are stored.

See: newS3Location smart constructor.

Constructors

S3Location'

Fields

bucketName :: Maybe Text

The Amazon S3 bucket name where the source code files provided with the project request are stored.

bucketKey :: Maybe Text

The Amazon S3 object key where the source code files provided with the project request are stored.

newS3Location :: S3Location

Create a value of S3Location with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bucketName:S3Location', s3Location_bucketName - The Amazon S3 bucket name where the source code files provided with the project request are stored.

$sel:bucketKey:S3Location', s3Location_bucketKey - The Amazon S3 object key where the source code files provided with the project request are stored.

s3Location_bucketName :: Lens' S3Location (Maybe Text)

The Amazon S3 bucket name where the source code files provided with the project request are stored.

s3Location_bucketKey :: Lens' S3Location (Maybe Text)

The Amazon S3 object key where the source code files provided with the project request are stored.

TeamMember

data TeamMember

Information about a team member in a project.

See: newTeamMember smart constructor.

Constructors

TeamMember'

Fields

remoteAccessAllowed :: Maybe Bool

Whether the user is allowed to remotely access project resources using an SSH public/private key pair.

userArn :: Text

The Amazon Resource Name (ARN) of the user in IAM.

projectRole :: Text

The role assigned to the user in the project. Project roles have different levels of access. For more information, see Working with Teams in the AWS CodeStar User Guide.

newTeamMember

Arguments

:: Text
$sel:userArn:TeamMember'

-> Text
$sel:projectRole:TeamMember'

-> TeamMember

Create a value of TeamMember with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:remoteAccessAllowed:TeamMember', teamMember_remoteAccessAllowed - Whether the user is allowed to remotely access project resources using an SSH public/private key pair.

$sel:userArn:TeamMember', teamMember_userArn - The Amazon Resource Name (ARN) of the user in IAM.

$sel:projectRole:TeamMember', teamMember_projectRole - The role assigned to the user in the project. Project roles have different levels of access. For more information, see Working with Teams in the AWS CodeStar User Guide.

teamMember_remoteAccessAllowed :: Lens' TeamMember (Maybe Bool)

Whether the user is allowed to remotely access project resources using an SSH public/private key pair.

teamMember_userArn :: Lens' TeamMember Text

The Amazon Resource Name (ARN) of the user in IAM.

teamMember_projectRole :: Lens' TeamMember Text

The role assigned to the user in the project. Project roles have different levels of access. For more information, see Working with Teams in the AWS CodeStar User Guide.

Toolchain

data Toolchain

The toolchain template file provided with the project request. AWS CodeStar uses the template to provision the toolchain stack in AWS CloudFormation.

See: newToolchain smart constructor.

Constructors

Toolchain'

Fields

stackParameters :: Maybe (HashMap Text (Sensitive Text))

The list of parameter overrides to be passed into the toolchain template during stack provisioning, if any.

roleArn :: Maybe Text

The service role ARN for AWS CodeStar to use for the toolchain template during stack provisioning.

source :: ToolchainSource

The Amazon S3 location where the toolchain template file provided with the project request is stored. AWS CodeStar retrieves the file during project creation.

newToolchain

Arguments

:: ToolchainSource
$sel:source:Toolchain'

-> Toolchain

Create a value of Toolchain with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stackParameters:Toolchain', toolchain_stackParameters - The list of parameter overrides to be passed into the toolchain template during stack provisioning, if any.

$sel:roleArn:Toolchain', toolchain_roleArn - The service role ARN for AWS CodeStar to use for the toolchain template during stack provisioning.

$sel:source:Toolchain', toolchain_source - The Amazon S3 location where the toolchain template file provided with the project request is stored. AWS CodeStar retrieves the file during project creation.

toolchain_stackParameters :: Lens' Toolchain (Maybe (HashMap Text Text))

The list of parameter overrides to be passed into the toolchain template during stack provisioning, if any.

toolchain_roleArn :: Lens' Toolchain (Maybe Text)

The service role ARN for AWS CodeStar to use for the toolchain template during stack provisioning.

toolchain_source :: Lens' Toolchain ToolchainSource

The Amazon S3 location where the toolchain template file provided with the project request is stored. AWS CodeStar retrieves the file during project creation.

ToolchainSource

data ToolchainSource

The Amazon S3 location where the toolchain template file provided with the project request is stored. AWS CodeStar retrieves the file during project creation.

See: newToolchainSource smart constructor.

Constructors

ToolchainSource'

Fields

s3 :: S3Location

The Amazon S3 bucket where the toolchain template file provided with the project request is stored.

newToolchainSource

Arguments

:: S3Location
$sel:s3:ToolchainSource'

-> ToolchainSource

Create a value of ToolchainSource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3:ToolchainSource', toolchainSource_s3 - The Amazon S3 bucket where the toolchain template file provided with the project request is stored.

toolchainSource_s3 :: Lens' ToolchainSource S3Location

The Amazon S3 bucket where the toolchain template file provided with the project request is stored.

UserProfileSummary

data UserProfileSummary

Information about a user's profile in AWS CodeStar.

See: newUserProfileSummary smart constructor.

Constructors

UserProfileSummary'

Fields

userArn :: Maybe Text

The Amazon Resource Name (ARN) of the user in IAM.

sshPublicKey :: Maybe Text

The SSH public key associated with the user in AWS CodeStar. If a project owner allows the user remote access to project resources, this public key will be used along with the user's private key for SSH access.

displayName :: Maybe (Sensitive Text)

The display name of a user in AWS CodeStar. For example, this could be set to both first and last name ("Mary Major") or a single name ("Mary"). The display name is also used to generate the initial icon associated with the user in AWS CodeStar projects. If spaces are included in the display name, the first character that appears after the space will be used as the second character in the user initial icon. The initial icon displays a maximum of two characters, so a display name with more than one space (for example "Mary Jane Major") would generate an initial icon using the first character and the first character after the space ("MJ", not "MM").

emailAddress :: Maybe (Sensitive Text)

The email address associated with the user.

newUserProfileSummary :: UserProfileSummary

Create a value of UserProfileSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userArn:UserProfileSummary', userProfileSummary_userArn - The Amazon Resource Name (ARN) of the user in IAM.

$sel:sshPublicKey:UserProfileSummary', userProfileSummary_sshPublicKey - The SSH public key associated with the user in AWS CodeStar. If a project owner allows the user remote access to project resources, this public key will be used along with the user's private key for SSH access.

$sel:displayName:UserProfileSummary', userProfileSummary_displayName - The display name of a user in AWS CodeStar. For example, this could be set to both first and last name ("Mary Major") or a single name ("Mary"). The display name is also used to generate the initial icon associated with the user in AWS CodeStar projects. If spaces are included in the display name, the first character that appears after the space will be used as the second character in the user initial icon. The initial icon displays a maximum of two characters, so a display name with more than one space (for example "Mary Jane Major") would generate an initial icon using the first character and the first character after the space ("MJ", not "MM").

$sel:emailAddress:UserProfileSummary', userProfileSummary_emailAddress - The email address associated with the user.

userProfileSummary_userArn :: Lens' UserProfileSummary (Maybe Text)

The Amazon Resource Name (ARN) of the user in IAM.

userProfileSummary_sshPublicKey :: Lens' UserProfileSummary (Maybe Text)

The SSH public key associated with the user in AWS CodeStar. If a project owner allows the user remote access to project resources, this public key will be used along with the user's private key for SSH access.

userProfileSummary_displayName :: Lens' UserProfileSummary (Maybe Text)

The display name of a user in AWS CodeStar. For example, this could be set to both first and last name ("Mary Major") or a single name ("Mary"). The display name is also used to generate the initial icon associated with the user in AWS CodeStar projects. If spaces are included in the display name, the first character that appears after the space will be used as the second character in the user initial icon. The initial icon displays a maximum of two characters, so a display name with more than one space (for example "Mary Jane Major") would generate an initial icon using the first character and the first character after the space ("MJ", not "MM").

userProfileSummary_emailAddress :: Lens' UserProfileSummary (Maybe Text)

The email address associated with the user.