Network.AWS.CodeStar.Types.TeamMember

Description

Documentation

data TeamMember

Information about a team member in a project.

See: newTeamMember smart constructor.

Constructors

TeamMember'

Fields

remoteAccessAllowed :: Maybe Bool

Whether the user is allowed to remotely access project resources using an SSH public/private key pair.

userArn :: Text

The Amazon Resource Name (ARN) of the user in IAM.

projectRole :: Text

The role assigned to the user in the project. Project roles have different levels of access. For more information, see Working with Teams in the AWS CodeStar User Guide.

newTeamMember

Arguments

:: Text
$sel:userArn:TeamMember'

-> Text
$sel:projectRole:TeamMember'

-> TeamMember

Create a value of TeamMember with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:remoteAccessAllowed:TeamMember', teamMember_remoteAccessAllowed - Whether the user is allowed to remotely access project resources using an SSH public/private key pair.

$sel:userArn:TeamMember', teamMember_userArn - The Amazon Resource Name (ARN) of the user in IAM.

$sel:projectRole:TeamMember', teamMember_projectRole - The role assigned to the user in the project. Project roles have different levels of access. For more information, see Working with Teams in the AWS CodeStar User Guide.

teamMember_remoteAccessAllowed :: Lens' TeamMember (Maybe Bool)

Whether the user is allowed to remotely access project resources using an SSH public/private key pair.

teamMember_userArn :: Lens' TeamMember Text

The Amazon Resource Name (ARN) of the user in IAM.

teamMember_projectRole :: Lens' TeamMember Text

The role assigned to the user in the project. Project roles have different levels of access. For more information, see Working with Teams in the AWS CodeStar User Guide.