Network.AWS.CodeStar.Types.UserProfileSummary

Description

Documentation

data UserProfileSummary

Information about a user's profile in AWS CodeStar.

See: newUserProfileSummary smart constructor.

Constructors

UserProfileSummary'

Fields

sshPublicKey :: Maybe Text

The SSH public key associated with the user in AWS CodeStar. If a project owner allows the user remote access to project resources, this public key will be used along with the user's private key for SSH access.

userArn :: Maybe Text

The Amazon Resource Name (ARN) of the user in IAM.

emailAddress :: Maybe (Sensitive Text)

The email address associated with the user.

displayName :: Maybe (Sensitive Text)

The display name of a user in AWS CodeStar. For example, this could be set to both first and last name ("Mary Major") or a single name ("Mary"). The display name is also used to generate the initial icon associated with the user in AWS CodeStar projects. If spaces are included in the display name, the first character that appears after the space will be used as the second character in the user initial icon. The initial icon displays a maximum of two characters, so a display name with more than one space (for example "Mary Jane Major") would generate an initial icon using the first character and the first character after the space ("MJ", not "MM").

newUserProfileSummary :: UserProfileSummary

Create a value of UserProfileSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sshPublicKey:UserProfileSummary', userProfileSummary_sshPublicKey - The SSH public key associated with the user in AWS CodeStar. If a project owner allows the user remote access to project resources, this public key will be used along with the user's private key for SSH access.

$sel:userArn:UserProfileSummary', userProfileSummary_userArn - The Amazon Resource Name (ARN) of the user in IAM.

$sel:emailAddress:UserProfileSummary', userProfileSummary_emailAddress - The email address associated with the user.

$sel:displayName:UserProfileSummary', userProfileSummary_displayName - The display name of a user in AWS CodeStar. For example, this could be set to both first and last name ("Mary Major") or a single name ("Mary"). The display name is also used to generate the initial icon associated with the user in AWS CodeStar projects. If spaces are included in the display name, the first character that appears after the space will be used as the second character in the user initial icon. The initial icon displays a maximum of two characters, so a display name with more than one space (for example "Mary Jane Major") would generate an initial icon using the first character and the first character after the space ("MJ", not "MM").

userProfileSummary_sshPublicKey :: Lens' UserProfileSummary (Maybe Text)

The SSH public key associated with the user in AWS CodeStar. If a project owner allows the user remote access to project resources, this public key will be used along with the user's private key for SSH access.

userProfileSummary_userArn :: Lens' UserProfileSummary (Maybe Text)

The Amazon Resource Name (ARN) of the user in IAM.

userProfileSummary_emailAddress :: Lens' UserProfileSummary (Maybe Text)

The email address associated with the user.

userProfileSummary_displayName :: Lens' UserProfileSummary (Maybe Text)

The display name of a user in AWS CodeStar. For example, this could be set to both first and last name ("Mary Major") or a single name ("Mary"). The display name is also used to generate the initial icon associated with the user in AWS CodeStar projects. If spaces are included in the display name, the first character that appears after the space will be used as the second character in the user initial icon. The initial icon displays a maximum of two characters, so a display name with more than one space (for example "Mary Jane Major") would generate an initial icon using the first character and the first character after the space ("MJ", not "MM").