Network.AWS.CodeStar.UpdateTeamMember

Description

Updates a team member's attributes in an AWS CodeStar project. For example, you can change a team member's role in the project, or change whether they have remote access to project resources.

Creating a Request

data UpdateTeamMember

See: newUpdateTeamMember smart constructor.

Constructors

UpdateTeamMember'

Fields

remoteAccessAllowed :: Maybe Bool

Whether a team member is allowed to remotely access project resources using the SSH public key associated with the user's profile. Even if this is set to True, the user must associate a public key with their profile before the user can access resources.

projectRole :: Maybe Text

The role assigned to the user in the project. Project roles have different levels of access. For more information, see Working with Teams in the AWS CodeStar User Guide.

projectId :: Text

The ID of the project.

userArn :: Text

The Amazon Resource Name (ARN) of the user for whom you want to change team membership attributes.

newUpdateTeamMember

Arguments

:: Text
$sel:projectId:UpdateTeamMember'

-> Text
$sel:userArn:UpdateTeamMember'

-> UpdateTeamMember

Create a value of UpdateTeamMember with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:remoteAccessAllowed:UpdateTeamMember', updateTeamMember_remoteAccessAllowed - Whether a team member is allowed to remotely access project resources using the SSH public key associated with the user's profile. Even if this is set to True, the user must associate a public key with their profile before the user can access resources.

$sel:projectRole:UpdateTeamMember', updateTeamMember_projectRole - The role assigned to the user in the project. Project roles have different levels of access. For more information, see Working with Teams in the AWS CodeStar User Guide.

$sel:projectId:UpdateTeamMember', updateTeamMember_projectId - The ID of the project.

$sel:userArn:UpdateTeamMember', updateTeamMember_userArn - The Amazon Resource Name (ARN) of the user for whom you want to change team membership attributes.

Request Lenses

updateTeamMember_remoteAccessAllowed :: Lens' UpdateTeamMember (Maybe Bool)

Whether a team member is allowed to remotely access project resources using the SSH public key associated with the user's profile. Even if this is set to True, the user must associate a public key with their profile before the user can access resources.

updateTeamMember_projectRole :: Lens' UpdateTeamMember (Maybe Text)

The role assigned to the user in the project. Project roles have different levels of access. For more information, see Working with Teams in the AWS CodeStar User Guide.

updateTeamMember_projectId :: Lens' UpdateTeamMember Text

The ID of the project.

updateTeamMember_userArn :: Lens' UpdateTeamMember Text

The Amazon Resource Name (ARN) of the user for whom you want to change team membership attributes.

Destructuring the Response

data UpdateTeamMemberResponse

See: newUpdateTeamMemberResponse smart constructor.

Constructors

UpdateTeamMemberResponse'

Fields

userArn :: Maybe Text

The Amazon Resource Name (ARN) of the user whose team membership attributes were updated.

remoteAccessAllowed :: Maybe Bool

Whether a team member is allowed to remotely access project resources using the SSH public key associated with the user's profile.

projectRole :: Maybe Text

The project role granted to the user.

httpStatus :: Int

The response's http status code.

newUpdateTeamMemberResponse

Arguments

:: Int
$sel:httpStatus:UpdateTeamMemberResponse'

-> UpdateTeamMemberResponse

Create a value of UpdateTeamMemberResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userArn:UpdateTeamMember', updateTeamMemberResponse_userArn - The Amazon Resource Name (ARN) of the user whose team membership attributes were updated.

$sel:remoteAccessAllowed:UpdateTeamMember', updateTeamMemberResponse_remoteAccessAllowed - Whether a team member is allowed to remotely access project resources using the SSH public key associated with the user's profile.

$sel:projectRole:UpdateTeamMember', updateTeamMemberResponse_projectRole - The project role granted to the user.

$sel:httpStatus:UpdateTeamMemberResponse', updateTeamMemberResponse_httpStatus - The response's http status code.

Response Lenses

updateTeamMemberResponse_userArn :: Lens' UpdateTeamMemberResponse (Maybe Text)

The Amazon Resource Name (ARN) of the user whose team membership attributes were updated.

updateTeamMemberResponse_remoteAccessAllowed :: Lens' UpdateTeamMemberResponse (Maybe Bool)

Whether a team member is allowed to remotely access project resources using the SSH public key associated with the user's profile.

updateTeamMemberResponse_projectRole :: Lens' UpdateTeamMemberResponse (Maybe Text)

The project role granted to the user.

updateTeamMemberResponse_httpStatus :: Lens' UpdateTeamMemberResponse Int

The response's http status code.