Network.AWS.CodeStar.UpdateUserProfile

Description

Updates a user's profile in AWS CodeStar. The user profile is not project-specific. Information in the user profile is displayed wherever the user's information appears to other users in AWS CodeStar.

Creating a Request

data UpdateUserProfile

See: newUpdateUserProfile smart constructor.

Constructors

UpdateUserProfile'

Fields

sshPublicKey :: Maybe Text

The SSH public key associated with the user in AWS CodeStar. If a project owner allows the user remote access to project resources, this public key will be used along with the user's private key for SSH access.

displayName :: Maybe (Sensitive Text)

The name that is displayed as the friendly name for the user in AWS CodeStar.

emailAddress :: Maybe (Sensitive Text)

The email address that is displayed as part of the user's profile in AWS CodeStar.

userArn :: Text

The name that will be displayed as the friendly name for the user in AWS CodeStar.

newUpdateUserProfile

Arguments

:: Text
$sel:userArn:UpdateUserProfile'

-> UpdateUserProfile

Create a value of UpdateUserProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sshPublicKey:UpdateUserProfile', updateUserProfile_sshPublicKey - The SSH public key associated with the user in AWS CodeStar. If a project owner allows the user remote access to project resources, this public key will be used along with the user's private key for SSH access.

$sel:displayName:UpdateUserProfile', updateUserProfile_displayName - The name that is displayed as the friendly name for the user in AWS CodeStar.

$sel:emailAddress:UpdateUserProfile', updateUserProfile_emailAddress - The email address that is displayed as part of the user's profile in AWS CodeStar.

$sel:userArn:UpdateUserProfile', updateUserProfile_userArn - The name that will be displayed as the friendly name for the user in AWS CodeStar.

Request Lenses

updateUserProfile_sshPublicKey :: Lens' UpdateUserProfile (Maybe Text)

The SSH public key associated with the user in AWS CodeStar. If a project owner allows the user remote access to project resources, this public key will be used along with the user's private key for SSH access.

updateUserProfile_displayName :: Lens' UpdateUserProfile (Maybe Text)

The name that is displayed as the friendly name for the user in AWS CodeStar.

updateUserProfile_emailAddress :: Lens' UpdateUserProfile (Maybe Text)

The email address that is displayed as part of the user's profile in AWS CodeStar.

updateUserProfile_userArn :: Lens' UpdateUserProfile Text

The name that will be displayed as the friendly name for the user in AWS CodeStar.

Destructuring the Response

data UpdateUserProfileResponse

See: newUpdateUserProfileResponse smart constructor.

Constructors

UpdateUserProfileResponse'

Fields

createdTimestamp :: Maybe POSIX

The date the user profile was created, in timestamp format.

lastModifiedTimestamp :: Maybe POSIX

The date the user profile was last modified, in timestamp format.

sshPublicKey :: Maybe Text

The SSH public key associated with the user in AWS CodeStar. This is the public portion of the public/private keypair the user can use to access project resources if a project owner allows the user remote access to those resources.

displayName :: Maybe (Sensitive Text)

The name that is displayed as the friendly name for the user in AWS CodeStar.

emailAddress :: Maybe (Sensitive Text)

The email address that is displayed as part of the user's profile in AWS CodeStar.

httpStatus :: Int

The response's http status code.

userArn :: Text

The Amazon Resource Name (ARN) of the user in IAM.

newUpdateUserProfileResponse

Arguments

:: Int
$sel:httpStatus:UpdateUserProfileResponse'

-> Text
$sel:userArn:UpdateUserProfile'

-> UpdateUserProfileResponse

Create a value of UpdateUserProfileResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createdTimestamp:UpdateUserProfileResponse', updateUserProfileResponse_createdTimestamp - The date the user profile was created, in timestamp format.

$sel:lastModifiedTimestamp:UpdateUserProfileResponse', updateUserProfileResponse_lastModifiedTimestamp - The date the user profile was last modified, in timestamp format.

$sel:sshPublicKey:UpdateUserProfile', updateUserProfileResponse_sshPublicKey - The SSH public key associated with the user in AWS CodeStar. This is the public portion of the public/private keypair the user can use to access project resources if a project owner allows the user remote access to those resources.

$sel:displayName:UpdateUserProfile', updateUserProfileResponse_displayName - The name that is displayed as the friendly name for the user in AWS CodeStar.

$sel:emailAddress:UpdateUserProfile', updateUserProfileResponse_emailAddress - The email address that is displayed as part of the user's profile in AWS CodeStar.

$sel:httpStatus:UpdateUserProfileResponse', updateUserProfileResponse_httpStatus - The response's http status code.

$sel:userArn:UpdateUserProfile', updateUserProfileResponse_userArn - The Amazon Resource Name (ARN) of the user in IAM.

Response Lenses

updateUserProfileResponse_createdTimestamp :: Lens' UpdateUserProfileResponse (Maybe UTCTime)

The date the user profile was created, in timestamp format.

updateUserProfileResponse_lastModifiedTimestamp :: Lens' UpdateUserProfileResponse (Maybe UTCTime)

The date the user profile was last modified, in timestamp format.

updateUserProfileResponse_sshPublicKey :: Lens' UpdateUserProfileResponse (Maybe Text)

The SSH public key associated with the user in AWS CodeStar. This is the public portion of the public/private keypair the user can use to access project resources if a project owner allows the user remote access to those resources.

updateUserProfileResponse_displayName :: Lens' UpdateUserProfileResponse (Maybe Text)

The name that is displayed as the friendly name for the user in AWS CodeStar.

updateUserProfileResponse_emailAddress :: Lens' UpdateUserProfileResponse (Maybe Text)

The email address that is displayed as part of the user's profile in AWS CodeStar.

updateUserProfileResponse_httpStatus :: Lens' UpdateUserProfileResponse Int

The response's http status code.

updateUserProfileResponse_userArn :: Lens' UpdateUserProfileResponse Text

The Amazon Resource Name (ARN) of the user in IAM.