Network.AWS.CodeStarConnections.GetHost

Description

Returns the host ARN and details such as status, provider type, endpoint, and, if applicable, the VPC configuration.

Creating a Request

data GetHost

See: newGetHost smart constructor.

Constructors

GetHost'

Fields

hostArn :: Text

The Amazon Resource Name (ARN) of the requested host.

newGetHost

Arguments

:: Text
$sel:hostArn:GetHost'

-> GetHost

Create a value of GetHost with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hostArn:GetHost', getHost_hostArn - The Amazon Resource Name (ARN) of the requested host.

Request Lenses

getHost_hostArn :: Lens' GetHost Text

The Amazon Resource Name (ARN) of the requested host.

Destructuring the Response

data GetHostResponse

See: newGetHostResponse smart constructor.

Constructors

GetHostResponse'

Fields

status :: Maybe Text

The status of the requested host.

providerEndpoint :: Maybe Text

The endpoint of the infrastructure represented by the requested host.

providerType :: Maybe ProviderType

The provider type of the requested host, such as GitHub Enterprise Server.

name :: Maybe Text

The name of the requested host.

vpcConfiguration :: Maybe VpcConfiguration

The VPC configuration of the requested host.

httpStatus :: Int

The response's http status code.

newGetHostResponse

Arguments

:: Int
$sel:httpStatus:GetHostResponse'

-> GetHostResponse

Create a value of GetHostResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:GetHostResponse', getHostResponse_status - The status of the requested host.

$sel:providerEndpoint:GetHostResponse', getHostResponse_providerEndpoint - The endpoint of the infrastructure represented by the requested host.

$sel:providerType:GetHostResponse', getHostResponse_providerType - The provider type of the requested host, such as GitHub Enterprise Server.

$sel:name:GetHostResponse', getHostResponse_name - The name of the requested host.

$sel:vpcConfiguration:GetHostResponse', getHostResponse_vpcConfiguration - The VPC configuration of the requested host.

$sel:httpStatus:GetHostResponse', getHostResponse_httpStatus - The response's http status code.

Response Lenses

getHostResponse_status :: Lens' GetHostResponse (Maybe Text)

The status of the requested host.

getHostResponse_providerEndpoint :: Lens' GetHostResponse (Maybe Text)

The endpoint of the infrastructure represented by the requested host.

getHostResponse_providerType :: Lens' GetHostResponse (Maybe ProviderType)

The provider type of the requested host, such as GitHub Enterprise Server.

getHostResponse_name :: Lens' GetHostResponse (Maybe Text)

The name of the requested host.

getHostResponse_vpcConfiguration :: Lens' GetHostResponse (Maybe VpcConfiguration)

The VPC configuration of the requested host.

getHostResponse_httpStatus :: Lens' GetHostResponse Int

The response's http status code.