Network.AWS.CodeStarConnections.Lens

Description

Operations

CreateHost

createHost_vpcConfiguration :: Lens' CreateHost (Maybe VpcConfiguration)

The VPC configuration to be provisioned for the host. A VPC must be configured and the infrastructure to be represented by the host must already be connected to the VPC.

createHost_tags :: Lens' CreateHost (Maybe [Tag])

Undocumented member.

createHost_name :: Lens' CreateHost Text

The name of the host to be created. The name must be unique in the calling AWS account.

createHost_providerType :: Lens' CreateHost ProviderType

The name of the installed provider to be associated with your connection. The host resource represents the infrastructure where your provider type is installed. The valid provider type is GitHub Enterprise Server.

createHost_providerEndpoint :: Lens' CreateHost Text

The endpoint of the infrastructure to be represented by the host after it is created.

createHostResponse_hostArn :: Lens' CreateHostResponse (Maybe Text)

The Amazon Resource Name (ARN) of the host to be created.

createHostResponse_tags :: Lens' CreateHostResponse (Maybe [Tag])

Undocumented member.

createHostResponse_httpStatus :: Lens' CreateHostResponse Int

The response's http status code.

ListConnections

listConnections_hostArnFilter :: Lens' ListConnections (Maybe Text)

Filters the list of connections to those associated with a specified host.

listConnections_nextToken :: Lens' ListConnections (Maybe Text)

The token that was returned from the previous ListConnections call, which can be used to return the next set of connections in the list.

listConnections_maxResults :: Lens' ListConnections (Maybe Natural)

The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned nextToken value.

listConnections_providerTypeFilter :: Lens' ListConnections (Maybe ProviderType)

Filters the list of connections to those associated with a specified provider, such as Bitbucket.

listConnectionsResponse_connections :: Lens' ListConnectionsResponse (Maybe [Connection])

A list of connections and the details for each connection, such as status, owner, and provider type.

listConnectionsResponse_nextToken :: Lens' ListConnectionsResponse (Maybe Text)

A token that can be used in the next ListConnections call. To view all items in the list, continue to call this operation with each subsequent token until no more nextToken values are returned.

listConnectionsResponse_httpStatus :: Lens' ListConnectionsResponse Int

The response's http status code.

DeleteConnection

deleteConnection_connectionArn :: Lens' DeleteConnection Text

The Amazon Resource Name (ARN) of the connection to be deleted.

The ARN is never reused if the connection is deleted.

deleteConnectionResponse_httpStatus :: Lens' DeleteConnectionResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The Amazon Resource Name (ARN) of the resource for which you want to get information about tags, if any.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe [Tag])

A list of tag key and value pairs associated with the specified resource.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

CreateConnection

createConnection_providerType :: Lens' CreateConnection (Maybe ProviderType)

The name of the external provider where your third-party code repository is configured.

createConnection_hostArn :: Lens' CreateConnection (Maybe Text)

The Amazon Resource Name (ARN) of the host associated with the connection to be created.

createConnection_tags :: Lens' CreateConnection (Maybe [Tag])

The key-value pair to use when tagging the resource.

createConnection_connectionName :: Lens' CreateConnection Text

The name of the connection to be created. The name must be unique in the calling AWS account.

createConnectionResponse_tags :: Lens' CreateConnectionResponse (Maybe [Tag])

Specifies the tags applied to the resource.

createConnectionResponse_httpStatus :: Lens' CreateConnectionResponse Int

The response's http status code.

createConnectionResponse_connectionArn :: Lens' CreateConnectionResponse Text

The Amazon Resource Name (ARN) of the connection to be created. The ARN is used as the connection reference when the connection is shared between AWS services.

The ARN is never reused if the connection is deleted.

GetConnection

getConnection_connectionArn :: Lens' GetConnection Text

The Amazon Resource Name (ARN) of a connection.

getConnectionResponse_connection :: Lens' GetConnectionResponse (Maybe Connection)

The connection details, such as status, owner, and provider type.

getConnectionResponse_httpStatus :: Lens' GetConnectionResponse Int

The response's http status code.

DeleteHost

deleteHost_hostArn :: Lens' DeleteHost Text

The Amazon Resource Name (ARN) of the host to be deleted.

deleteHostResponse_httpStatus :: Lens' DeleteHostResponse Int

The response's http status code.

UpdateHost

updateHost_providerEndpoint :: Lens' UpdateHost (Maybe Text)

The URL or endpoint of the host to be updated.

updateHost_vpcConfiguration :: Lens' UpdateHost (Maybe VpcConfiguration)

The VPC configuration of the host to be updated. A VPC must be configured and the infrastructure to be represented by the host must already be connected to the VPC.

updateHost_hostArn :: Lens' UpdateHost Text

The Amazon Resource Name (ARN) of the host to be updated.

updateHostResponse_httpStatus :: Lens' UpdateHostResponse Int

The response's http status code.

ListHosts

listHosts_nextToken :: Lens' ListHosts (Maybe Text)

The token that was returned from the previous ListHosts call, which can be used to return the next set of hosts in the list.

listHosts_maxResults :: Lens' ListHosts (Maybe Natural)

The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned nextToken value.

listHostsResponse_hosts :: Lens' ListHostsResponse (Maybe [Host])

A list of hosts and the details for each host, such as status, endpoint, and provider type.

listHostsResponse_nextToken :: Lens' ListHostsResponse (Maybe Text)

A token that can be used in the next ListHosts call. To view all items in the list, continue to call this operation with each subsequent token until no more nextToken values are returned.

listHostsResponse_httpStatus :: Lens' ListHostsResponse Int

The response's http status code.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The Amazon Resource Name (ARN) of the resource to which you want to add or update tags.

tagResource_tags :: Lens' TagResource [Tag]

The tags you want to modify or add to the resource.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

UntagResource

untagResource_resourceArn :: Lens' UntagResource Text

The Amazon Resource Name (ARN) of the resource to remove tags from.

untagResource_tagKeys :: Lens' UntagResource [Text]

The list of keys for the tags to be removed from the resource.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

GetHost

getHost_hostArn :: Lens' GetHost Text

The Amazon Resource Name (ARN) of the requested host.

getHostResponse_status :: Lens' GetHostResponse (Maybe Text)

The status of the requested host.

getHostResponse_providerEndpoint :: Lens' GetHostResponse (Maybe Text)

The endpoint of the infrastructure represented by the requested host.

getHostResponse_providerType :: Lens' GetHostResponse (Maybe ProviderType)

The provider type of the requested host, such as GitHub Enterprise Server.

getHostResponse_name :: Lens' GetHostResponse (Maybe Text)

The name of the requested host.

getHostResponse_vpcConfiguration :: Lens' GetHostResponse (Maybe VpcConfiguration)

The VPC configuration of the requested host.

getHostResponse_httpStatus :: Lens' GetHostResponse Int

The response's http status code.

Types

Connection

connection_ownerAccountId :: Lens' Connection (Maybe Text)

The identifier of the external provider where your third-party code repository is configured. For Bitbucket, this is the account ID of the owner of the Bitbucket repository.

connection_providerType :: Lens' Connection (Maybe ProviderType)

The name of the external provider where your third-party code repository is configured.

connection_connectionName :: Lens' Connection (Maybe Text)

The name of the connection. Connection names must be unique in an AWS user account.

connection_connectionStatus :: Lens' Connection (Maybe ConnectionStatus)

The current status of the connection.

connection_hostArn :: Lens' Connection (Maybe Text)

The Amazon Resource Name (ARN) of the host associated with the connection.

connection_connectionArn :: Lens' Connection (Maybe Text)

The Amazon Resource Name (ARN) of the connection. The ARN is used as the connection reference when the connection is shared between AWS services.

The ARN is never reused if the connection is deleted.

Host

host_status :: Lens' Host (Maybe Text)

The status of the host, such as PENDING, AVAILABLE, VPC_CONFIG_DELETING, VPC_CONFIG_INITIALIZING, and VPC_CONFIG_FAILED_INITIALIZATION.

host_providerEndpoint :: Lens' Host (Maybe Text)

The endpoint of the infrastructure where your provider type is installed.

host_providerType :: Lens' Host (Maybe ProviderType)

The name of the installed provider to be associated with your connection. The host resource represents the infrastructure where your provider type is installed. The valid provider type is GitHub Enterprise Server.

host_statusMessage :: Lens' Host (Maybe Text)

The status description for the host.

host_name :: Lens' Host (Maybe Text)

The name of the host.

host_hostArn :: Lens' Host (Maybe Text)

The Amazon Resource Name (ARN) of the host.

host_vpcConfiguration :: Lens' Host (Maybe VpcConfiguration)

The VPC configuration provisioned for the host.

Tag

tag_key :: Lens' Tag Text

The tag's key.

tag_value :: Lens' Tag Text

The tag's value.

VpcConfiguration

vpcConfiguration_tlsCertificate :: Lens' VpcConfiguration (Maybe Text)

The value of the Transport Layer Security (TLS) certificate associated with the infrastructure where your provider type is installed.

vpcConfiguration_vpcId :: Lens' VpcConfiguration Text

The ID of the Amazon VPC connected to the infrastructure where your provider type is installed.

vpcConfiguration_subnetIds :: Lens' VpcConfiguration (NonEmpty Text)

The ID of the subnet or subnets associated with the Amazon VPC connected to the infrastructure where your provider type is installed.

vpcConfiguration_securityGroupIds :: Lens' VpcConfiguration (NonEmpty Text)

The ID of the security group or security groups associated with the Amazon VPC connected to the infrastructure where your provider type is installed.