Network.AWS.CodeStarConnections.Types

Description

Service Configuration

defaultService :: Service

API version 2019-12-01 of the Amazon CodeStar connections SDK configuration.

Errors

_ResourceUnavailableException :: AsError a => Getting (First ServiceError) a ServiceError

Resource not found. Verify the ARN for the host resource and try again.

_UnsupportedOperationException :: AsError a => Getting (First ServiceError) a ServiceError

The operation is not supported. Check the connection status and try again.

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

Two conflicting operations have been made on the same resource.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

Resource not found. Verify the connection resource ARN and try again.

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Exceeded the maximum limit for connections.

ConnectionStatus

newtype ConnectionStatus

Constructors

ConnectionStatus'

Fields

fromConnectionStatus :: Text

Bundled Patterns

pattern ConnectionStatus_AVAILABLE :: ConnectionStatus

pattern ConnectionStatus_ERROR :: ConnectionStatus

pattern ConnectionStatus_PENDING :: ConnectionStatus

ProviderType

newtype ProviderType

Constructors

ProviderType'

Fields

fromProviderType :: Text

Bundled Patterns

pattern ProviderType_Bitbucket :: ProviderType

pattern ProviderType_GitHub :: ProviderType

pattern ProviderType_GitHubEnterpriseServer :: ProviderType

Connection

data Connection

A resource that is used to connect third-party source providers with services like AWS CodePipeline.

Note: A connection created through CloudFormation, the CLI, or the SDK is in `PENDING` status by default. You can make its status `AVAILABLE` by updating the connection in the console.

See: newConnection smart constructor.

Constructors

Connection'

Fields

ownerAccountId :: Maybe Text

The identifier of the external provider where your third-party code repository is configured. For Bitbucket, this is the account ID of the owner of the Bitbucket repository.

providerType :: Maybe ProviderType

The name of the external provider where your third-party code repository is configured.

connectionName :: Maybe Text

The name of the connection. Connection names must be unique in an AWS user account.

connectionStatus :: Maybe ConnectionStatus

The current status of the connection.

hostArn :: Maybe Text

The Amazon Resource Name (ARN) of the host associated with the connection.

connectionArn :: Maybe Text

The Amazon Resource Name (ARN) of the connection. The ARN is used as the connection reference when the connection is shared between AWS services.

The ARN is never reused if the connection is deleted.

newConnection :: Connection

Create a value of Connection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ownerAccountId:Connection', connection_ownerAccountId - The identifier of the external provider where your third-party code repository is configured. For Bitbucket, this is the account ID of the owner of the Bitbucket repository.

$sel:providerType:Connection', connection_providerType - The name of the external provider where your third-party code repository is configured.

$sel:connectionName:Connection', connection_connectionName - The name of the connection. Connection names must be unique in an AWS user account.

$sel:connectionStatus:Connection', connection_connectionStatus - The current status of the connection.

$sel:hostArn:Connection', connection_hostArn - The Amazon Resource Name (ARN) of the host associated with the connection.

$sel:connectionArn:Connection', connection_connectionArn - The Amazon Resource Name (ARN) of the connection. The ARN is used as the connection reference when the connection is shared between AWS services.

The ARN is never reused if the connection is deleted.

connection_ownerAccountId :: Lens' Connection (Maybe Text)

The identifier of the external provider where your third-party code repository is configured. For Bitbucket, this is the account ID of the owner of the Bitbucket repository.

connection_providerType :: Lens' Connection (Maybe ProviderType)

The name of the external provider where your third-party code repository is configured.

connection_connectionName :: Lens' Connection (Maybe Text)

The name of the connection. Connection names must be unique in an AWS user account.

connection_connectionStatus :: Lens' Connection (Maybe ConnectionStatus)

The current status of the connection.

connection_hostArn :: Lens' Connection (Maybe Text)

The Amazon Resource Name (ARN) of the host associated with the connection.

connection_connectionArn :: Lens' Connection (Maybe Text)

The Amazon Resource Name (ARN) of the connection. The ARN is used as the connection reference when the connection is shared between AWS services.

The ARN is never reused if the connection is deleted.

Host

data Host

A resource that represents the infrastructure where a third-party provider is installed. The host is used when you create connections to an installed third-party provider type, such as GitHub Enterprise Server. You create one host for all connections to that provider.

A host created through the CLI or the SDK is in `PENDING` status by default. You can make its status `AVAILABLE` by setting up the host in the console.

See: newHost smart constructor.

Constructors

Host'

Fields

status :: Maybe Text

The status of the host, such as PENDING, AVAILABLE, VPC_CONFIG_DELETING, VPC_CONFIG_INITIALIZING, and VPC_CONFIG_FAILED_INITIALIZATION.

providerEndpoint :: Maybe Text

The endpoint of the infrastructure where your provider type is installed.

providerType :: Maybe ProviderType

The name of the installed provider to be associated with your connection. The host resource represents the infrastructure where your provider type is installed. The valid provider type is GitHub Enterprise Server.

statusMessage :: Maybe Text

The status description for the host.

name :: Maybe Text

The name of the host.

hostArn :: Maybe Text

The Amazon Resource Name (ARN) of the host.

vpcConfiguration :: Maybe VpcConfiguration

The VPC configuration provisioned for the host.

newHost :: Host

Create a value of Host with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Host', host_status - The status of the host, such as PENDING, AVAILABLE, VPC_CONFIG_DELETING, VPC_CONFIG_INITIALIZING, and VPC_CONFIG_FAILED_INITIALIZATION.

$sel:providerEndpoint:Host', host_providerEndpoint - The endpoint of the infrastructure where your provider type is installed.

$sel:providerType:Host', host_providerType - The name of the installed provider to be associated with your connection. The host resource represents the infrastructure where your provider type is installed. The valid provider type is GitHub Enterprise Server.

$sel:statusMessage:Host', host_statusMessage - The status description for the host.

$sel:name:Host', host_name - The name of the host.

$sel:hostArn:Host', host_hostArn - The Amazon Resource Name (ARN) of the host.

$sel:vpcConfiguration:Host', host_vpcConfiguration - The VPC configuration provisioned for the host.

host_status :: Lens' Host (Maybe Text)

The status of the host, such as PENDING, AVAILABLE, VPC_CONFIG_DELETING, VPC_CONFIG_INITIALIZING, and VPC_CONFIG_FAILED_INITIALIZATION.

host_providerEndpoint :: Lens' Host (Maybe Text)

The endpoint of the infrastructure where your provider type is installed.

host_providerType :: Lens' Host (Maybe ProviderType)

The name of the installed provider to be associated with your connection. The host resource represents the infrastructure where your provider type is installed. The valid provider type is GitHub Enterprise Server.

host_statusMessage :: Lens' Host (Maybe Text)

The status description for the host.

host_name :: Lens' Host (Maybe Text)

The name of the host.

host_hostArn :: Lens' Host (Maybe Text)

The Amazon Resource Name (ARN) of the host.

host_vpcConfiguration :: Lens' Host (Maybe VpcConfiguration)

The VPC configuration provisioned for the host.

Tag

data Tag

A tag is a key-value pair that is used to manage the resource.

This tag is available for use by AWS services that support tags.

See: newTag smart constructor.

Constructors

Tag'

Fields

key :: Text

The tag's key.

value :: Text

The tag's value.

newTag

Arguments

:: Text
$sel:key:Tag'

-> Text
$sel:value:Tag'

-> Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:Tag', tag_key - The tag's key.

$sel:value:Tag', tag_value - The tag's value.

tag_key :: Lens' Tag Text

The tag's key.

tag_value :: Lens' Tag Text

The tag's value.

VpcConfiguration

data VpcConfiguration

The VPC configuration provisioned for the host.

See: newVpcConfiguration smart constructor.

Constructors

VpcConfiguration'

Fields

tlsCertificate :: Maybe Text

The value of the Transport Layer Security (TLS) certificate associated with the infrastructure where your provider type is installed.

vpcId :: Text

The ID of the Amazon VPC connected to the infrastructure where your provider type is installed.

subnetIds :: NonEmpty Text

The ID of the subnet or subnets associated with the Amazon VPC connected to the infrastructure where your provider type is installed.

securityGroupIds :: NonEmpty Text

The ID of the security group or security groups associated with the Amazon VPC connected to the infrastructure where your provider type is installed.

newVpcConfiguration

Arguments

:: Text
$sel:vpcId:VpcConfiguration'

-> NonEmpty Text
$sel:subnetIds:VpcConfiguration'

-> NonEmpty Text
$sel:securityGroupIds:VpcConfiguration'

-> VpcConfiguration

Create a value of VpcConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tlsCertificate:VpcConfiguration', vpcConfiguration_tlsCertificate - The value of the Transport Layer Security (TLS) certificate associated with the infrastructure where your provider type is installed.

$sel:vpcId:VpcConfiguration', vpcConfiguration_vpcId - The ID of the Amazon VPC connected to the infrastructure where your provider type is installed.

$sel:subnetIds:VpcConfiguration', vpcConfiguration_subnetIds - The ID of the subnet or subnets associated with the Amazon VPC connected to the infrastructure where your provider type is installed.

$sel:securityGroupIds:VpcConfiguration', vpcConfiguration_securityGroupIds - The ID of the security group or security groups associated with the Amazon VPC connected to the infrastructure where your provider type is installed.

vpcConfiguration_tlsCertificate :: Lens' VpcConfiguration (Maybe Text)

The value of the Transport Layer Security (TLS) certificate associated with the infrastructure where your provider type is installed.

vpcConfiguration_vpcId :: Lens' VpcConfiguration Text

The ID of the Amazon VPC connected to the infrastructure where your provider type is installed.

vpcConfiguration_subnetIds :: Lens' VpcConfiguration (NonEmpty Text)

The ID of the subnet or subnets associated with the Amazon VPC connected to the infrastructure where your provider type is installed.

vpcConfiguration_securityGroupIds :: Lens' VpcConfiguration (NonEmpty Text)

The ID of the security group or security groups associated with the Amazon VPC connected to the infrastructure where your provider type is installed.