Network.AWS.CodeStarConnections.Types.Host

Description

Documentation

data Host

A resource that represents the infrastructure where a third-party provider is installed. The host is used when you create connections to an installed third-party provider type, such as GitHub Enterprise Server. You create one host for all connections to that provider.

A host created through the CLI or the SDK is in `PENDING` status by default. You can make its status `AVAILABLE` by setting up the host in the console.

See: newHost smart constructor.

Constructors

Host'

Fields

status :: Maybe Text

The status of the host, such as PENDING, AVAILABLE, VPC_CONFIG_DELETING, VPC_CONFIG_INITIALIZING, and VPC_CONFIG_FAILED_INITIALIZATION.

providerEndpoint :: Maybe Text

The endpoint of the infrastructure where your provider type is installed.

providerType :: Maybe ProviderType

The name of the installed provider to be associated with your connection. The host resource represents the infrastructure where your provider type is installed. The valid provider type is GitHub Enterprise Server.

statusMessage :: Maybe Text

The status description for the host.

name :: Maybe Text

The name of the host.

hostArn :: Maybe Text

The Amazon Resource Name (ARN) of the host.

vpcConfiguration :: Maybe VpcConfiguration

The VPC configuration provisioned for the host.

newHost :: Host

Create a value of Host with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Host', host_status - The status of the host, such as PENDING, AVAILABLE, VPC_CONFIG_DELETING, VPC_CONFIG_INITIALIZING, and VPC_CONFIG_FAILED_INITIALIZATION.

$sel:providerEndpoint:Host', host_providerEndpoint - The endpoint of the infrastructure where your provider type is installed.

$sel:providerType:Host', host_providerType - The name of the installed provider to be associated with your connection. The host resource represents the infrastructure where your provider type is installed. The valid provider type is GitHub Enterprise Server.

$sel:statusMessage:Host', host_statusMessage - The status description for the host.

$sel:name:Host', host_name - The name of the host.

$sel:hostArn:Host', host_hostArn - The Amazon Resource Name (ARN) of the host.

$sel:vpcConfiguration:Host', host_vpcConfiguration - The VPC configuration provisioned for the host.

host_status :: Lens' Host (Maybe Text)

The status of the host, such as PENDING, AVAILABLE, VPC_CONFIG_DELETING, VPC_CONFIG_INITIALIZING, and VPC_CONFIG_FAILED_INITIALIZATION.

host_providerEndpoint :: Lens' Host (Maybe Text)

The endpoint of the infrastructure where your provider type is installed.

host_providerType :: Lens' Host (Maybe ProviderType)

The name of the installed provider to be associated with your connection. The host resource represents the infrastructure where your provider type is installed. The valid provider type is GitHub Enterprise Server.

host_statusMessage :: Lens' Host (Maybe Text)

The status description for the host.

host_name :: Lens' Host (Maybe Text)

The name of the host.

host_hostArn :: Lens' Host (Maybe Text)

The Amazon Resource Name (ARN) of the host.

host_vpcConfiguration :: Lens' Host (Maybe VpcConfiguration)

The VPC configuration provisioned for the host.