Network.AWS.CodeStarNotifications

Description

Derived from API version 2019-10-15 of the AWS service descriptions, licensed under Apache 2.0.

This AWS CodeStar Notifications API Reference provides descriptions and usage examples of the operations and data types for the AWS CodeStar Notifications API. You can use the AWS CodeStar Notifications API to work with the following objects:

Notification rules, by calling the following:

CreateNotificationRule, which creates a notification rule for a resource in your account.

DeleteNotificationRule, which deletes a notification rule.

DescribeNotificationRule, which provides information about a notification rule.

ListNotificationRules, which lists the notification rules associated with your account.

UpdateNotificationRule, which changes the name, events, or targets associated with a notification rule.

Subscribe, which subscribes a target to a notification rule.

Unsubscribe, which removes a target from a notification rule.

Targets, by calling the following:

DeleteTarget, which removes a notification rule target (SNS topic) from a notification rule.

ListTargets, which lists the targets associated with a notification rule.

Events, by calling the following:

ListEventTypes, which lists the event types you can include in a notification rule.

Tags, by calling the following:

ListTagsForResource, which lists the tags already associated with a notification rule in your account.

TagResource, which associates a tag you provide with a notification rule in your account.

UntagResource, which removes a tag from a notification rule in your account.

For information about how to use AWS CodeStar Notifications, see link in the CodeStarNotifications User Guide.

Service Configuration

defaultService :: Service

API version 2019-10-15 of the Amazon CodeStar Notifications SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by CodeStarNotifications.

ValidationException

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

One or more parameter values are not valid.

AccessDeniedException

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

AWS CodeStar Notifications can't create the notification rule because you do not have sufficient permissions.

ResourceAlreadyExistsException

_ResourceAlreadyExistsException :: AsError a => Getting (First ServiceError) a ServiceError

A resource with the same name or ID already exists. Notification rule names must be unique in your AWS account.

ConfigurationException

_ConfigurationException :: AsError a => Getting (First ServiceError) a ServiceError

Some or all of the configuration is incomplete, missing, or not valid.

ConcurrentModificationException

_ConcurrentModificationException :: AsError a => Getting (First ServiceError) a ServiceError

AWS CodeStar Notifications can't complete the request because the resource is being modified by another process. Wait a few minutes and try again.

InvalidNextTokenException

_InvalidNextTokenException :: AsError a => Getting (First ServiceError) a ServiceError

The value for the enumeration token used in the request to return the next batch of the results is not valid.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

AWS CodeStar Notifications can't find a resource that matches the provided ARN.

LimitExceededException

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

One of the AWS CodeStar Notifications limits has been exceeded. Limits apply to accounts, notification rules, notifications, resources, and targets. For more information, see Limits.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

CreateNotificationRule

data CreateNotificationRule

See: newCreateNotificationRule smart constructor.

Constructors

CreateNotificationRule' (Maybe NotificationRuleStatus) (Maybe Text) (Maybe (HashMap Text Text)) (Sensitive Text) [Text] Text [Target] DetailType

newCreateNotificationRule

Arguments

:: Text
$sel:name:CreateNotificationRule'

-> Text
$sel:resource:CreateNotificationRule'

-> DetailType
$sel:detailType:CreateNotificationRule'

-> CreateNotificationRule

Create a value of CreateNotificationRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:CreateNotificationRule', createNotificationRule_status - The status of the notification rule. The default value is ENABLED. If the status is set to DISABLED, notifications aren't sent for the notification rule.

$sel:clientRequestToken:CreateNotificationRule', createNotificationRule_clientRequestToken - A unique, client-generated idempotency token that, when provided in a request, ensures the request cannot be repeated with a changed parameter. If a request with the same parameters is received and a token is included, the request returns information about the initial request that used that token.

The AWS SDKs prepopulate client request tokens. If you are using an AWS SDK, an idempotency token is created for you.

$sel:tags:CreateNotificationRule', createNotificationRule_tags - A list of tags to apply to this notification rule. Key names cannot start with "aws".

$sel:name:CreateNotificationRule', createNotificationRule_name - The name for the notification rule. Notifictaion rule names must be unique in your AWS account.

$sel:eventTypeIds:CreateNotificationRule', createNotificationRule_eventTypeIds - A list of event types associated with this notification rule. For a list of allowed events, see EventTypeSummary.

$sel:resource:CreateNotificationRule', createNotificationRule_resource - The Amazon Resource Name (ARN) of the resource to associate with the notification rule. Supported resources include pipelines in AWS CodePipeline, repositories in AWS CodeCommit, and build projects in AWS CodeBuild.

$sel:targets:CreateNotificationRule', createNotificationRule_targets - A list of Amazon Resource Names (ARNs) of SNS topics to associate with the notification rule.

$sel:detailType:CreateNotificationRule', createNotificationRule_detailType - The level of detail to include in the notifications for this resource. BASIC will include only the contents of the event as it would appear in AWS CloudWatch. FULL will include any supplemental information provided by AWS CodeStar Notifications and/or the service for the resource for which the notification is created.

data CreateNotificationRuleResponse

See: newCreateNotificationRuleResponse smart constructor.

Constructors

CreateNotificationRuleResponse' (Maybe Text) Int

newCreateNotificationRuleResponse

Arguments

:: Int
$sel:httpStatus:CreateNotificationRuleResponse'

-> CreateNotificationRuleResponse

Create a value of CreateNotificationRuleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:CreateNotificationRuleResponse', createNotificationRuleResponse_arn - The Amazon Resource Name (ARN) of the notification rule.

$sel:httpStatus:CreateNotificationRuleResponse', createNotificationRuleResponse_httpStatus - The response's http status code.

UpdateNotificationRule

data UpdateNotificationRule

See: newUpdateNotificationRule smart constructor.

Constructors

UpdateNotificationRule' (Maybe NotificationRuleStatus) (Maybe [Text]) (Maybe DetailType) (Maybe (Sensitive Text)) (Maybe [Target]) Text

newUpdateNotificationRule

Arguments

:: Text
$sel:arn:UpdateNotificationRule'

-> UpdateNotificationRule

Create a value of UpdateNotificationRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:UpdateNotificationRule', updateNotificationRule_status - The status of the notification rule. Valid statuses include enabled (sending notifications) or disabled (not sending notifications).

$sel:eventTypeIds:UpdateNotificationRule', updateNotificationRule_eventTypeIds - A list of event types associated with this notification rule.

$sel:detailType:UpdateNotificationRule', updateNotificationRule_detailType - The level of detail to include in the notifications for this resource. BASIC will include only the contents of the event as it would appear in AWS CloudWatch. FULL will include any supplemental information provided by AWS CodeStar Notifications and/or the service for the resource for which the notification is created.

$sel:name:UpdateNotificationRule', updateNotificationRule_name - The name of the notification rule.

$sel:targets:UpdateNotificationRule', updateNotificationRule_targets - The address and type of the targets to receive notifications from this notification rule.

$sel:arn:UpdateNotificationRule', updateNotificationRule_arn - The Amazon Resource Name (ARN) of the notification rule.

data UpdateNotificationRuleResponse

See: newUpdateNotificationRuleResponse smart constructor.

Constructors

UpdateNotificationRuleResponse' Int

newUpdateNotificationRuleResponse

Arguments

:: Int
$sel:httpStatus:UpdateNotificationRuleResponse'

-> UpdateNotificationRuleResponse

Create a value of UpdateNotificationRuleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateNotificationRuleResponse', updateNotificationRuleResponse_httpStatus - The response's http status code.

DeleteNotificationRule

data DeleteNotificationRule

See: newDeleteNotificationRule smart constructor.

Constructors

DeleteNotificationRule' Text

newDeleteNotificationRule

Arguments

:: Text
$sel:arn:DeleteNotificationRule'

-> DeleteNotificationRule

Create a value of DeleteNotificationRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:DeleteNotificationRule', deleteNotificationRule_arn - The Amazon Resource Name (ARN) of the notification rule you want to delete.

data DeleteNotificationRuleResponse

See: newDeleteNotificationRuleResponse smart constructor.

Constructors

DeleteNotificationRuleResponse' (Maybe Text) Int

newDeleteNotificationRuleResponse

Arguments

:: Int
$sel:httpStatus:DeleteNotificationRuleResponse'

-> DeleteNotificationRuleResponse

Create a value of DeleteNotificationRuleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:DeleteNotificationRule', deleteNotificationRuleResponse_arn - The Amazon Resource Name (ARN) of the deleted notification rule.

$sel:httpStatus:DeleteNotificationRuleResponse', deleteNotificationRuleResponse_httpStatus - The response's http status code.

ListTagsForResource

data ListTagsForResource

See: newListTagsForResource smart constructor.

Constructors

ListTagsForResource' Text

newListTagsForResource

Arguments

:: Text
$sel:arn:ListTagsForResource'

-> ListTagsForResource

Create a value of ListTagsForResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:ListTagsForResource', listTagsForResource_arn - The Amazon Resource Name (ARN) for the notification rule.

data ListTagsForResourceResponse

See: newListTagsForResourceResponse smart constructor.

Constructors

ListTagsForResourceResponse' (Maybe (HashMap Text Text)) Int

newListTagsForResourceResponse

Arguments

:: Int
$sel:httpStatus:ListTagsForResourceResponse'

-> ListTagsForResourceResponse

Create a value of ListTagsForResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:ListTagsForResourceResponse', listTagsForResourceResponse_tags - The tags associated with the notification rule.

$sel:httpStatus:ListTagsForResourceResponse', listTagsForResourceResponse_httpStatus - The response's http status code.

ListEventTypes (Paginated)

data ListEventTypes

See: newListEventTypes smart constructor.

Constructors

ListEventTypes' (Maybe [ListEventTypesFilter]) (Maybe Text) (Maybe Natural)

newListEventTypes :: ListEventTypes

Create a value of ListEventTypes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:ListEventTypes', listEventTypes_filters - The filters to use to return information by service or resource type.

$sel:nextToken:ListEventTypes', listEventTypes_nextToken - An enumeration token that, when provided in a request, returns the next batch of the results.

$sel:maxResults:ListEventTypes', listEventTypes_maxResults - A non-negative integer used to limit the number of returned results. The default number is 50. The maximum number of results that can be returned is 100.

data ListEventTypesResponse

See: newListEventTypesResponse smart constructor.

Constructors

ListEventTypesResponse' (Maybe [EventTypeSummary]) (Maybe Text) Int

newListEventTypesResponse

Arguments

:: Int
$sel:httpStatus:ListEventTypesResponse'

-> ListEventTypesResponse

Create a value of ListEventTypesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventTypes:ListEventTypesResponse', listEventTypesResponse_eventTypes - Information about each event, including service name, resource type, event ID, and event name.

$sel:nextToken:ListEventTypes', listEventTypesResponse_nextToken - An enumeration token that can be used in a request to return the next batch of the results.

$sel:httpStatus:ListEventTypesResponse', listEventTypesResponse_httpStatus - The response's http status code.

DeleteTarget

data DeleteTarget

See: newDeleteTarget smart constructor.

Constructors

DeleteTarget' (Maybe Bool) (Sensitive Text)

newDeleteTarget

Arguments

:: Text
$sel:targetAddress:DeleteTarget'

-> DeleteTarget

Create a value of DeleteTarget with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:forceUnsubscribeAll:DeleteTarget', deleteTarget_forceUnsubscribeAll - A Boolean value that can be used to delete all associations with this SNS topic. The default value is FALSE. If set to TRUE, all associations between that target and every notification rule in your AWS account are deleted.

$sel:targetAddress:DeleteTarget', deleteTarget_targetAddress - The Amazon Resource Name (ARN) of the SNS topic to delete.

data DeleteTargetResponse

See: newDeleteTargetResponse smart constructor.

Constructors

DeleteTargetResponse' Int

newDeleteTargetResponse

Arguments

:: Int
$sel:httpStatus:DeleteTargetResponse'

-> DeleteTargetResponse

Create a value of DeleteTargetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteTargetResponse', deleteTargetResponse_httpStatus - The response's http status code.

ListNotificationRules (Paginated)

data ListNotificationRules

See: newListNotificationRules smart constructor.

Constructors

ListNotificationRules' (Maybe [ListNotificationRulesFilter]) (Maybe Text) (Maybe Natural)

newListNotificationRules :: ListNotificationRules

Create a value of ListNotificationRules with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:ListNotificationRules', listNotificationRules_filters - The filters to use to return information by service or resource type. For valid values, see ListNotificationRulesFilter.

A filter with the same name can appear more than once when used with OR statements. Filters with different names should be applied with AND statements.

$sel:nextToken:ListNotificationRules', listNotificationRules_nextToken - An enumeration token that, when provided in a request, returns the next batch of the results.

$sel:maxResults:ListNotificationRules', listNotificationRules_maxResults - A non-negative integer used to limit the number of returned results. The maximum number of results that can be returned is 100.

data ListNotificationRulesResponse

See: newListNotificationRulesResponse smart constructor.

Constructors

ListNotificationRulesResponse' (Maybe Text) (Maybe [NotificationRuleSummary]) Int

newListNotificationRulesResponse

Arguments

:: Int
$sel:httpStatus:ListNotificationRulesResponse'

-> ListNotificationRulesResponse

Create a value of ListNotificationRulesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListNotificationRules', listNotificationRulesResponse_nextToken - An enumeration token that can be used in a request to return the next batch of the results.

$sel:notificationRules:ListNotificationRulesResponse', listNotificationRulesResponse_notificationRules - The list of notification rules for the AWS account, by Amazon Resource Name (ARN) and ID.

$sel:httpStatus:ListNotificationRulesResponse', listNotificationRulesResponse_httpStatus - The response's http status code.

ListTargets (Paginated)

data ListTargets

See: newListTargets smart constructor.

Constructors

ListTargets' (Maybe [ListTargetsFilter]) (Maybe Text) (Maybe Natural)

newListTargets :: ListTargets

Create a value of ListTargets with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:ListTargets', listTargets_filters - The filters to use to return information by service or resource type. Valid filters include target type, target address, and target status.

A filter with the same name can appear more than once when used with OR statements. Filters with different names should be applied with AND statements.

$sel:nextToken:ListTargets', listTargets_nextToken - An enumeration token that, when provided in a request, returns the next batch of the results.

$sel:maxResults:ListTargets', listTargets_maxResults - A non-negative integer used to limit the number of returned results. The maximum number of results that can be returned is 100.

data ListTargetsResponse

See: newListTargetsResponse smart constructor.

Constructors

ListTargetsResponse' (Maybe Text) (Maybe [TargetSummary]) Int

newListTargetsResponse

Arguments

:: Int
$sel:httpStatus:ListTargetsResponse'

-> ListTargetsResponse

Create a value of ListTargetsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListTargets', listTargetsResponse_nextToken - An enumeration token that can be used in a request to return the next batch of results.

$sel:targets:ListTargetsResponse', listTargetsResponse_targets - The list of notification rule targets.

$sel:httpStatus:ListTargetsResponse', listTargetsResponse_httpStatus - The response's http status code.

TagResource

data TagResource

See: newTagResource smart constructor.

Constructors

TagResource' Text (HashMap Text Text)

newTagResource

Arguments

:: Text
$sel:arn:TagResource'

-> TagResource

Create a value of TagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:TagResource', tagResource_arn - The Amazon Resource Name (ARN) of the notification rule to tag.

$sel:tags:TagResource', tagResource_tags - The list of tags to associate with the resource. Tag key names cannot start with "aws".

data TagResourceResponse

See: newTagResourceResponse smart constructor.

Constructors

TagResourceResponse' (Maybe (HashMap Text Text)) Int

newTagResourceResponse

Arguments

:: Int
$sel:httpStatus:TagResourceResponse'

-> TagResourceResponse

Create a value of TagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:TagResource', tagResourceResponse_tags - The list of tags associated with the resource.

$sel:httpStatus:TagResourceResponse', tagResourceResponse_httpStatus - The response's http status code.

Subscribe

data Subscribe

See: newSubscribe smart constructor.

Constructors

Subscribe' (Maybe Text) Text Target

newSubscribe

Arguments

:: Text
$sel:arn:Subscribe'

-> Target
$sel:target:Subscribe'

-> Subscribe

Create a value of Subscribe with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientRequestToken:Subscribe', subscribe_clientRequestToken - An enumeration token that, when provided in a request, returns the next batch of the results.

$sel:arn:Subscribe', subscribe_arn - The Amazon Resource Name (ARN) of the notification rule for which you want to create the association.

$sel:target:Subscribe', subscribe_target - Undocumented member.

data SubscribeResponse

See: newSubscribeResponse smart constructor.

Constructors

SubscribeResponse' (Maybe Text) Int

newSubscribeResponse

Arguments

:: Int
$sel:httpStatus:SubscribeResponse'

-> SubscribeResponse

Create a value of SubscribeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:Subscribe', subscribeResponse_arn - The Amazon Resource Name (ARN) of the notification rule for which you have created assocations.

$sel:httpStatus:SubscribeResponse', subscribeResponse_httpStatus - The response's http status code.

UntagResource

data UntagResource

See: newUntagResource smart constructor.

Constructors

UntagResource' Text [Text]

newUntagResource

Arguments

:: Text
$sel:arn:UntagResource'

-> UntagResource

Create a value of UntagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:UntagResource', untagResource_arn - The Amazon Resource Name (ARN) of the notification rule from which to remove the tags.

$sel:tagKeys:UntagResource', untagResource_tagKeys - The key names of the tags to remove.

data UntagResourceResponse

See: newUntagResourceResponse smart constructor.

Constructors

UntagResourceResponse' Int

newUntagResourceResponse

Arguments

:: Int
$sel:httpStatus:UntagResourceResponse'

-> UntagResourceResponse

Create a value of UntagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UntagResourceResponse', untagResourceResponse_httpStatus - The response's http status code.

Unsubscribe

data Unsubscribe

See: newUnsubscribe smart constructor.

Constructors

Unsubscribe' Text (Sensitive Text)

newUnsubscribe

Arguments

:: Text
$sel:arn:Unsubscribe'

-> Text
$sel:targetAddress:Unsubscribe'

-> Unsubscribe

Create a value of Unsubscribe with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:Unsubscribe', unsubscribe_arn - The Amazon Resource Name (ARN) of the notification rule.

$sel:targetAddress:Unsubscribe', unsubscribe_targetAddress - The ARN of the SNS topic to unsubscribe from the notification rule.

data UnsubscribeResponse

See: newUnsubscribeResponse smart constructor.

Constructors

UnsubscribeResponse' Int Text

newUnsubscribeResponse

Arguments

:: Int
$sel:httpStatus:UnsubscribeResponse'

-> Text
$sel:arn:Unsubscribe'

-> UnsubscribeResponse

Create a value of UnsubscribeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UnsubscribeResponse', unsubscribeResponse_httpStatus - The response's http status code.

$sel:arn:Unsubscribe', unsubscribeResponse_arn - The Amazon Resource Name (ARN) of the the notification rule from which you have removed a subscription.

DescribeNotificationRule

data DescribeNotificationRule

See: newDescribeNotificationRule smart constructor.

Constructors

DescribeNotificationRule' Text

newDescribeNotificationRule

Arguments

:: Text
$sel:arn:DescribeNotificationRule'

-> DescribeNotificationRule

Create a value of DescribeNotificationRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:DescribeNotificationRule', describeNotificationRule_arn - The Amazon Resource Name (ARN) of the notification rule.

data DescribeNotificationRuleResponse

See: newDescribeNotificationRuleResponse smart constructor.

Constructors

DescribeNotificationRuleResponse' (Maybe NotificationRuleStatus) (Maybe [EventTypeSummary]) (Maybe POSIX) (Maybe Text) (Maybe DetailType) (Maybe (Sensitive Text)) (Maybe [TargetSummary]) (Maybe Text) (Maybe POSIX) (Maybe (HashMap Text Text)) Int Text

newDescribeNotificationRuleResponse

Arguments

:: Int
$sel:httpStatus:DescribeNotificationRuleResponse'

-> Text
$sel:arn:DescribeNotificationRule'

-> DescribeNotificationRuleResponse

Create a value of DescribeNotificationRuleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DescribeNotificationRuleResponse', describeNotificationRuleResponse_status - The status of the notification rule. Valid statuses are on (sending notifications) or off (not sending notifications).

$sel:eventTypes:DescribeNotificationRuleResponse', describeNotificationRuleResponse_eventTypes - A list of the event types associated with the notification rule.

$sel:lastModifiedTimestamp:DescribeNotificationRuleResponse', describeNotificationRuleResponse_lastModifiedTimestamp - The date and time the notification rule was most recently updated, in timestamp format.

$sel:createdBy:DescribeNotificationRuleResponse', describeNotificationRuleResponse_createdBy - The name or email alias of the person who created the notification rule.

$sel:detailType:DescribeNotificationRuleResponse', describeNotificationRuleResponse_detailType - The level of detail included in the notifications for this resource. BASIC will include only the contents of the event as it would appear in AWS CloudWatch. FULL will include any supplemental information provided by AWS CodeStar Notifications and/or the service for the resource for which the notification is created.

$sel:name:DescribeNotificationRuleResponse', describeNotificationRuleResponse_name - The name of the notification rule.

$sel:targets:DescribeNotificationRuleResponse', describeNotificationRuleResponse_targets - A list of the SNS topics associated with the notification rule.

$sel:resource:DescribeNotificationRuleResponse', describeNotificationRuleResponse_resource - The Amazon Resource Name (ARN) of the resource associated with the notification rule.

$sel:createdTimestamp:DescribeNotificationRuleResponse', describeNotificationRuleResponse_createdTimestamp - The date and time the notification rule was created, in timestamp format.

$sel:tags:DescribeNotificationRuleResponse', describeNotificationRuleResponse_tags - The tags associated with the notification rule.

$sel:httpStatus:DescribeNotificationRuleResponse', describeNotificationRuleResponse_httpStatus - The response's http status code.

$sel:arn:DescribeNotificationRule', describeNotificationRuleResponse_arn - The Amazon Resource Name (ARN) of the notification rule.

Types

DetailType

newtype DetailType

Constructors

DetailType'

Fields

fromDetailType :: Text

Bundled Patterns

pattern DetailType_BASIC :: DetailType

pattern DetailType_FULL :: DetailType

ListEventTypesFilterName

newtype ListEventTypesFilterName

Constructors

ListEventTypesFilterName'

Fields

fromListEventTypesFilterName :: Text

Bundled Patterns

pattern ListEventTypesFilterName_RESOURCE_TYPE :: ListEventTypesFilterName

pattern ListEventTypesFilterName_SERVICE_NAME :: ListEventTypesFilterName

ListNotificationRulesFilterName

newtype ListNotificationRulesFilterName

Constructors

ListNotificationRulesFilterName'

Fields

fromListNotificationRulesFilterName :: Text

Bundled Patterns

pattern ListNotificationRulesFilterName_CREATED_BY :: ListNotificationRulesFilterName

pattern ListNotificationRulesFilterName_EVENT_TYPE_ID :: ListNotificationRulesFilterName

pattern ListNotificationRulesFilterName_RESOURCE :: ListNotificationRulesFilterName

pattern ListNotificationRulesFilterName_TARGET_ADDRESS :: ListNotificationRulesFilterName

ListTargetsFilterName

newtype ListTargetsFilterName

Constructors

ListTargetsFilterName'

Fields

fromListTargetsFilterName :: Text

Bundled Patterns

pattern ListTargetsFilterName_TARGET_ADDRESS :: ListTargetsFilterName

pattern ListTargetsFilterName_TARGET_STATUS :: ListTargetsFilterName

pattern ListTargetsFilterName_TARGET_TYPE :: ListTargetsFilterName

NotificationRuleStatus

newtype NotificationRuleStatus

Constructors

NotificationRuleStatus'

Fields

fromNotificationRuleStatus :: Text

Bundled Patterns

pattern NotificationRuleStatus_DISABLED :: NotificationRuleStatus

pattern NotificationRuleStatus_ENABLED :: NotificationRuleStatus

TargetStatus

newtype TargetStatus

Constructors

TargetStatus'

Fields

fromTargetStatus :: Text

Bundled Patterns

pattern TargetStatus_ACTIVE :: TargetStatus

pattern TargetStatus_DEACTIVATED :: TargetStatus

pattern TargetStatus_INACTIVE :: TargetStatus

pattern TargetStatus_PENDING :: TargetStatus

pattern TargetStatus_UNREACHABLE :: TargetStatus

EventTypeSummary

data EventTypeSummary

Returns information about an event that has triggered a notification rule.

See: newEventTypeSummary smart constructor.

Constructors

EventTypeSummary' (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text)

newEventTypeSummary :: EventTypeSummary

Create a value of EventTypeSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceType:EventTypeSummary', eventTypeSummary_resourceType - The resource type of the event.

$sel:eventTypeName:EventTypeSummary', eventTypeSummary_eventTypeName - The name of the event.

$sel:eventTypeId:EventTypeSummary', eventTypeSummary_eventTypeId - The system-generated ID of the event.

$sel:serviceName:EventTypeSummary', eventTypeSummary_serviceName - The name of the service for which the event applies.

ListEventTypesFilter

data ListEventTypesFilter

Information about a filter to apply to the list of returned event types. You can filter by resource type or service name.

See: newListEventTypesFilter smart constructor.

Constructors

ListEventTypesFilter' ListEventTypesFilterName Text

newListEventTypesFilter

Arguments

:: ListEventTypesFilterName
$sel:name:ListEventTypesFilter'

-> Text
$sel:value:ListEventTypesFilter'

-> ListEventTypesFilter

Create a value of ListEventTypesFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:ListEventTypesFilter', listEventTypesFilter_name - The system-generated name of the filter type you want to filter by.

$sel:value:ListEventTypesFilter', listEventTypesFilter_value - The name of the resource type (for example, pipeline) or service name (for example, CodePipeline) that you want to filter by.

ListNotificationRulesFilter

data ListNotificationRulesFilter

Information about a filter to apply to the list of returned notification rules. You can filter by event type, owner, resource, or target.

See: newListNotificationRulesFilter smart constructor.

Constructors

ListNotificationRulesFilter' ListNotificationRulesFilterName Text

newListNotificationRulesFilter

Arguments

:: ListNotificationRulesFilterName
$sel:name:ListNotificationRulesFilter'

-> Text
$sel:value:ListNotificationRulesFilter'

-> ListNotificationRulesFilter

Create a value of ListNotificationRulesFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:ListNotificationRulesFilter', listNotificationRulesFilter_name - The name of the attribute you want to use to filter the returned notification rules.

$sel:value:ListNotificationRulesFilter', listNotificationRulesFilter_value - The value of the attribute you want to use to filter the returned notification rules. For example, if you specify filtering by RESOURCE in Name, you might specify the ARN of a pipeline in AWS CodePipeline for the value.

ListTargetsFilter

data ListTargetsFilter

Information about a filter to apply to the list of returned targets. You can filter by target type, address, or status. For example, to filter results to notification rules that have active Amazon SNS topics as targets, you could specify a ListTargetsFilter Name as TargetType and a Value of SNS, and a Name of TARGET_STATUS and a Value of ACTIVE.

See: newListTargetsFilter smart constructor.

Constructors

ListTargetsFilter' ListTargetsFilterName Text

newListTargetsFilter

Arguments

:: ListTargetsFilterName
$sel:name:ListTargetsFilter'

-> Text
$sel:value:ListTargetsFilter'

-> ListTargetsFilter

Create a value of ListTargetsFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:ListTargetsFilter', listTargetsFilter_name - The name of the attribute you want to use to filter the returned targets.

$sel:value:ListTargetsFilter', listTargetsFilter_value - The value of the attribute you want to use to filter the returned targets. For example, if you specify SNS for the Target type, you could specify an Amazon Resource Name (ARN) for a topic as the value.

NotificationRuleSummary

data NotificationRuleSummary

Information about a specified notification rule.

See: newNotificationRuleSummary smart constructor.

Constructors

NotificationRuleSummary' (Maybe Text) (Maybe Text)

newNotificationRuleSummary :: NotificationRuleSummary

Create a value of NotificationRuleSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:NotificationRuleSummary', notificationRuleSummary_arn - The Amazon Resource Name (ARN) of the notification rule.

$sel:id:NotificationRuleSummary', notificationRuleSummary_id - The unique ID of the notification rule.

Target

data Target

Information about the SNS topics associated with a notification rule.

See: newTarget smart constructor.

Constructors

Target' (Maybe Text) (Maybe (Sensitive Text))

newTarget :: Target

Create a value of Target with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetType:Target', target_targetType - The target type. Can be an Amazon SNS topic.

$sel:targetAddress:Target', target_targetAddress - The Amazon Resource Name (ARN) of the SNS topic.

TargetSummary

data TargetSummary

Information about the targets specified for a notification rule.

See: newTargetSummary smart constructor.

Constructors

TargetSummary' (Maybe Text) (Maybe (Sensitive Text)) (Maybe TargetStatus)

newTargetSummary :: TargetSummary

Create a value of TargetSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetType:TargetSummary', targetSummary_targetType - The type of the target (for example, SNS).

$sel:targetAddress:TargetSummary', targetSummary_targetAddress - The Amazon Resource Name (ARN) of the SNS topic.

$sel:targetStatus:TargetSummary', targetSummary_targetStatus - The status of the target.