Network.AWS.CodeStarNotifications.CreateNotificationRule

Description

Creates a notification rule for a resource. The rule specifies the events you want notifications about and the targets (such as SNS topics) where you want to receive them.

Creating a Request

data CreateNotificationRule

See: newCreateNotificationRule smart constructor.

Constructors

CreateNotificationRule'

Fields

status :: Maybe NotificationRuleStatus

The status of the notification rule. The default value is ENABLED. If the status is set to DISABLED, notifications aren't sent for the notification rule.

clientRequestToken :: Maybe Text

A unique, client-generated idempotency token that, when provided in a request, ensures the request cannot be repeated with a changed parameter. If a request with the same parameters is received and a token is included, the request returns information about the initial request that used that token.

The AWS SDKs prepopulate client request tokens. If you are using an AWS SDK, an idempotency token is created for you.

tags :: Maybe (HashMap Text Text)

A list of tags to apply to this notification rule. Key names cannot start with "aws".

name :: Sensitive Text

The name for the notification rule. Notifictaion rule names must be unique in your AWS account.

eventTypeIds :: [Text]

A list of event types associated with this notification rule. For a list of allowed events, see EventTypeSummary.

resource :: Text

The Amazon Resource Name (ARN) of the resource to associate with the notification rule. Supported resources include pipelines in AWS CodePipeline, repositories in AWS CodeCommit, and build projects in AWS CodeBuild.

targets :: [Target]

A list of Amazon Resource Names (ARNs) of SNS topics to associate with the notification rule.

detailType :: DetailType

The level of detail to include in the notifications for this resource. BASIC will include only the contents of the event as it would appear in AWS CloudWatch. FULL will include any supplemental information provided by AWS CodeStar Notifications and/or the service for the resource for which the notification is created.

newCreateNotificationRule

Arguments

:: Text
$sel:name:CreateNotificationRule'

-> Text
$sel:resource:CreateNotificationRule'

-> DetailType
$sel:detailType:CreateNotificationRule'

-> CreateNotificationRule

Create a value of CreateNotificationRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:CreateNotificationRule', createNotificationRule_status - The status of the notification rule. The default value is ENABLED. If the status is set to DISABLED, notifications aren't sent for the notification rule.

$sel:clientRequestToken:CreateNotificationRule', createNotificationRule_clientRequestToken - A unique, client-generated idempotency token that, when provided in a request, ensures the request cannot be repeated with a changed parameter. If a request with the same parameters is received and a token is included, the request returns information about the initial request that used that token.

The AWS SDKs prepopulate client request tokens. If you are using an AWS SDK, an idempotency token is created for you.

$sel:tags:CreateNotificationRule', createNotificationRule_tags - A list of tags to apply to this notification rule. Key names cannot start with "aws".

$sel:name:CreateNotificationRule', createNotificationRule_name - The name for the notification rule. Notifictaion rule names must be unique in your AWS account.

$sel:eventTypeIds:CreateNotificationRule', createNotificationRule_eventTypeIds - A list of event types associated with this notification rule. For a list of allowed events, see EventTypeSummary.

$sel:resource:CreateNotificationRule', createNotificationRule_resource - The Amazon Resource Name (ARN) of the resource to associate with the notification rule. Supported resources include pipelines in AWS CodePipeline, repositories in AWS CodeCommit, and build projects in AWS CodeBuild.

$sel:targets:CreateNotificationRule', createNotificationRule_targets - A list of Amazon Resource Names (ARNs) of SNS topics to associate with the notification rule.

$sel:detailType:CreateNotificationRule', createNotificationRule_detailType - The level of detail to include in the notifications for this resource. BASIC will include only the contents of the event as it would appear in AWS CloudWatch. FULL will include any supplemental information provided by AWS CodeStar Notifications and/or the service for the resource for which the notification is created.

Request Lenses

createNotificationRule_status :: Lens' CreateNotificationRule (Maybe NotificationRuleStatus)

The status of the notification rule. The default value is ENABLED. If the status is set to DISABLED, notifications aren't sent for the notification rule.

createNotificationRule_clientRequestToken :: Lens' CreateNotificationRule (Maybe Text)

A unique, client-generated idempotency token that, when provided in a request, ensures the request cannot be repeated with a changed parameter. If a request with the same parameters is received and a token is included, the request returns information about the initial request that used that token.

The AWS SDKs prepopulate client request tokens. If you are using an AWS SDK, an idempotency token is created for you.

createNotificationRule_tags :: Lens' CreateNotificationRule (Maybe (HashMap Text Text))

A list of tags to apply to this notification rule. Key names cannot start with "aws".

createNotificationRule_name :: Lens' CreateNotificationRule Text

The name for the notification rule. Notifictaion rule names must be unique in your AWS account.

createNotificationRule_eventTypeIds :: Lens' CreateNotificationRule [Text]

A list of event types associated with this notification rule. For a list of allowed events, see EventTypeSummary.

createNotificationRule_resource :: Lens' CreateNotificationRule Text

The Amazon Resource Name (ARN) of the resource to associate with the notification rule. Supported resources include pipelines in AWS CodePipeline, repositories in AWS CodeCommit, and build projects in AWS CodeBuild.

createNotificationRule_targets :: Lens' CreateNotificationRule [Target]

A list of Amazon Resource Names (ARNs) of SNS topics to associate with the notification rule.

createNotificationRule_detailType :: Lens' CreateNotificationRule DetailType

The level of detail to include in the notifications for this resource. BASIC will include only the contents of the event as it would appear in AWS CloudWatch. FULL will include any supplemental information provided by AWS CodeStar Notifications and/or the service for the resource for which the notification is created.

Destructuring the Response

data CreateNotificationRuleResponse

See: newCreateNotificationRuleResponse smart constructor.

Constructors

CreateNotificationRuleResponse'

Fields

arn :: Maybe Text

The Amazon Resource Name (ARN) of the notification rule.

httpStatus :: Int

The response's http status code.

newCreateNotificationRuleResponse

Arguments

:: Int
$sel:httpStatus:CreateNotificationRuleResponse'

-> CreateNotificationRuleResponse

Create a value of CreateNotificationRuleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:CreateNotificationRuleResponse', createNotificationRuleResponse_arn - The Amazon Resource Name (ARN) of the notification rule.

$sel:httpStatus:CreateNotificationRuleResponse', createNotificationRuleResponse_httpStatus - The response's http status code.

Response Lenses

createNotificationRuleResponse_arn :: Lens' CreateNotificationRuleResponse (Maybe Text)

The Amazon Resource Name (ARN) of the notification rule.

createNotificationRuleResponse_httpStatus :: Lens' CreateNotificationRuleResponse Int

The response's http status code.