Network.AWS.CodeStarNotifications.DescribeNotificationRule

Description

Returns information about a specified notification rule.

Creating a Request

data DescribeNotificationRule

See: newDescribeNotificationRule smart constructor.

Constructors

DescribeNotificationRule'

Fields

arn :: Text

The Amazon Resource Name (ARN) of the notification rule.

newDescribeNotificationRule

Arguments

:: Text
$sel:arn:DescribeNotificationRule'

-> DescribeNotificationRule

Create a value of DescribeNotificationRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:DescribeNotificationRule', describeNotificationRule_arn - The Amazon Resource Name (ARN) of the notification rule.

Request Lenses

describeNotificationRule_arn :: Lens' DescribeNotificationRule Text

The Amazon Resource Name (ARN) of the notification rule.

Destructuring the Response

data DescribeNotificationRuleResponse

See: newDescribeNotificationRuleResponse smart constructor.

Constructors

DescribeNotificationRuleResponse'

Fields

status :: Maybe NotificationRuleStatus

The status of the notification rule. Valid statuses are on (sending notifications) or off (not sending notifications).

eventTypes :: Maybe [EventTypeSummary]

A list of the event types associated with the notification rule.

lastModifiedTimestamp :: Maybe POSIX

The date and time the notification rule was most recently updated, in timestamp format.

createdBy :: Maybe Text

The name or email alias of the person who created the notification rule.

detailType :: Maybe DetailType

The level of detail included in the notifications for this resource. BASIC will include only the contents of the event as it would appear in AWS CloudWatch. FULL will include any supplemental information provided by AWS CodeStar Notifications and/or the service for the resource for which the notification is created.

name :: Maybe (Sensitive Text)

The name of the notification rule.

targets :: Maybe [TargetSummary]

A list of the SNS topics associated with the notification rule.

resource :: Maybe Text

The Amazon Resource Name (ARN) of the resource associated with the notification rule.

createdTimestamp :: Maybe POSIX

The date and time the notification rule was created, in timestamp format.

tags :: Maybe (HashMap Text Text)

The tags associated with the notification rule.

httpStatus :: Int

The response's http status code.

arn :: Text

The Amazon Resource Name (ARN) of the notification rule.

newDescribeNotificationRuleResponse

Arguments

:: Int
$sel:httpStatus:DescribeNotificationRuleResponse'

-> Text
$sel:arn:DescribeNotificationRule'

-> DescribeNotificationRuleResponse

Create a value of DescribeNotificationRuleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DescribeNotificationRuleResponse', describeNotificationRuleResponse_status - The status of the notification rule. Valid statuses are on (sending notifications) or off (not sending notifications).

$sel:eventTypes:DescribeNotificationRuleResponse', describeNotificationRuleResponse_eventTypes - A list of the event types associated with the notification rule.

$sel:lastModifiedTimestamp:DescribeNotificationRuleResponse', describeNotificationRuleResponse_lastModifiedTimestamp - The date and time the notification rule was most recently updated, in timestamp format.

$sel:createdBy:DescribeNotificationRuleResponse', describeNotificationRuleResponse_createdBy - The name or email alias of the person who created the notification rule.

$sel:detailType:DescribeNotificationRuleResponse', describeNotificationRuleResponse_detailType - The level of detail included in the notifications for this resource. BASIC will include only the contents of the event as it would appear in AWS CloudWatch. FULL will include any supplemental information provided by AWS CodeStar Notifications and/or the service for the resource for which the notification is created.

$sel:name:DescribeNotificationRuleResponse', describeNotificationRuleResponse_name - The name of the notification rule.

$sel:targets:DescribeNotificationRuleResponse', describeNotificationRuleResponse_targets - A list of the SNS topics associated with the notification rule.

$sel:resource:DescribeNotificationRuleResponse', describeNotificationRuleResponse_resource - The Amazon Resource Name (ARN) of the resource associated with the notification rule.

$sel:createdTimestamp:DescribeNotificationRuleResponse', describeNotificationRuleResponse_createdTimestamp - The date and time the notification rule was created, in timestamp format.

$sel:tags:DescribeNotificationRuleResponse', describeNotificationRuleResponse_tags - The tags associated with the notification rule.

$sel:httpStatus:DescribeNotificationRuleResponse', describeNotificationRuleResponse_httpStatus - The response's http status code.

$sel:arn:DescribeNotificationRule', describeNotificationRuleResponse_arn - The Amazon Resource Name (ARN) of the notification rule.

Response Lenses

describeNotificationRuleResponse_status :: Lens' DescribeNotificationRuleResponse (Maybe NotificationRuleStatus)

The status of the notification rule. Valid statuses are on (sending notifications) or off (not sending notifications).

describeNotificationRuleResponse_eventTypes :: Lens' DescribeNotificationRuleResponse (Maybe [EventTypeSummary])

A list of the event types associated with the notification rule.

describeNotificationRuleResponse_lastModifiedTimestamp :: Lens' DescribeNotificationRuleResponse (Maybe UTCTime)

The date and time the notification rule was most recently updated, in timestamp format.

describeNotificationRuleResponse_createdBy :: Lens' DescribeNotificationRuleResponse (Maybe Text)

The name or email alias of the person who created the notification rule.

describeNotificationRuleResponse_detailType :: Lens' DescribeNotificationRuleResponse (Maybe DetailType)

The level of detail included in the notifications for this resource. BASIC will include only the contents of the event as it would appear in AWS CloudWatch. FULL will include any supplemental information provided by AWS CodeStar Notifications and/or the service for the resource for which the notification is created.

describeNotificationRuleResponse_name :: Lens' DescribeNotificationRuleResponse (Maybe Text)

The name of the notification rule.

describeNotificationRuleResponse_targets :: Lens' DescribeNotificationRuleResponse (Maybe [TargetSummary])

A list of the SNS topics associated with the notification rule.

describeNotificationRuleResponse_resource :: Lens' DescribeNotificationRuleResponse (Maybe Text)

The Amazon Resource Name (ARN) of the resource associated with the notification rule.

describeNotificationRuleResponse_createdTimestamp :: Lens' DescribeNotificationRuleResponse (Maybe UTCTime)

The date and time the notification rule was created, in timestamp format.

describeNotificationRuleResponse_tags :: Lens' DescribeNotificationRuleResponse (Maybe (HashMap Text Text))

The tags associated with the notification rule.

describeNotificationRuleResponse_httpStatus :: Lens' DescribeNotificationRuleResponse Int

The response's http status code.

describeNotificationRuleResponse_arn :: Lens' DescribeNotificationRuleResponse Text

The Amazon Resource Name (ARN) of the notification rule.