Network.AWS.CodeStarNotifications.ListEventTypes

Description

Returns information about the event types available for configuring notifications.

This operation returns paginated results.

Creating a Request

data ListEventTypes

See: newListEventTypes smart constructor.

Constructors

ListEventTypes'

Fields

filters :: Maybe [ListEventTypesFilter]

The filters to use to return information by service or resource type.

nextToken :: Maybe Text

An enumeration token that, when provided in a request, returns the next batch of the results.

maxResults :: Maybe Natural

A non-negative integer used to limit the number of returned results. The default number is 50. The maximum number of results that can be returned is 100.

newListEventTypes :: ListEventTypes

Create a value of ListEventTypes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:ListEventTypes', listEventTypes_filters - The filters to use to return information by service or resource type.

$sel:nextToken:ListEventTypes', listEventTypes_nextToken - An enumeration token that, when provided in a request, returns the next batch of the results.

$sel:maxResults:ListEventTypes', listEventTypes_maxResults - A non-negative integer used to limit the number of returned results. The default number is 50. The maximum number of results that can be returned is 100.

Request Lenses

listEventTypes_filters :: Lens' ListEventTypes (Maybe [ListEventTypesFilter])

The filters to use to return information by service or resource type.

listEventTypes_nextToken :: Lens' ListEventTypes (Maybe Text)

An enumeration token that, when provided in a request, returns the next batch of the results.

listEventTypes_maxResults :: Lens' ListEventTypes (Maybe Natural)

A non-negative integer used to limit the number of returned results. The default number is 50. The maximum number of results that can be returned is 100.

Destructuring the Response

data ListEventTypesResponse

See: newListEventTypesResponse smart constructor.

Constructors

ListEventTypesResponse'

Fields

eventTypes :: Maybe [EventTypeSummary]

Information about each event, including service name, resource type, event ID, and event name.

nextToken :: Maybe Text

An enumeration token that can be used in a request to return the next batch of the results.

httpStatus :: Int

The response's http status code.

newListEventTypesResponse

Arguments

:: Int
$sel:httpStatus:ListEventTypesResponse'

-> ListEventTypesResponse

Create a value of ListEventTypesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventTypes:ListEventTypesResponse', listEventTypesResponse_eventTypes - Information about each event, including service name, resource type, event ID, and event name.

$sel:nextToken:ListEventTypes', listEventTypesResponse_nextToken - An enumeration token that can be used in a request to return the next batch of the results.

$sel:httpStatus:ListEventTypesResponse', listEventTypesResponse_httpStatus - The response's http status code.

Response Lenses

listEventTypesResponse_eventTypes :: Lens' ListEventTypesResponse (Maybe [EventTypeSummary])

Information about each event, including service name, resource type, event ID, and event name.

listEventTypesResponse_nextToken :: Lens' ListEventTypesResponse (Maybe Text)

An enumeration token that can be used in a request to return the next batch of the results.

listEventTypesResponse_httpStatus :: Lens' ListEventTypesResponse Int

The response's http status code.