Network.AWS.CodeStarNotifications.Types

Description

Service Configuration

defaultService :: Service

API version 2019-10-15 of the Amazon CodeStar Notifications SDK configuration.

Errors

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

One or more parameter values are not valid.

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

AWS CodeStar Notifications can't create the notification rule because you do not have sufficient permissions.

_ResourceAlreadyExistsException :: AsError a => Getting (First ServiceError) a ServiceError

A resource with the same name or ID already exists. Notification rule names must be unique in your AWS account.

_ConfigurationException :: AsError a => Getting (First ServiceError) a ServiceError

Some or all of the configuration is incomplete, missing, or not valid.

_ConcurrentModificationException :: AsError a => Getting (First ServiceError) a ServiceError

AWS CodeStar Notifications can't complete the request because the resource is being modified by another process. Wait a few minutes and try again.

_InvalidNextTokenException :: AsError a => Getting (First ServiceError) a ServiceError

The value for the enumeration token used in the request to return the next batch of the results is not valid.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

AWS CodeStar Notifications can't find a resource that matches the provided ARN.

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

One of the AWS CodeStar Notifications limits has been exceeded. Limits apply to accounts, notification rules, notifications, resources, and targets. For more information, see Limits.

DetailType

newtype DetailType

Constructors

DetailType'

Fields

fromDetailType :: Text

Bundled Patterns

pattern DetailType_BASIC :: DetailType

pattern DetailType_FULL :: DetailType

ListEventTypesFilterName

newtype ListEventTypesFilterName

Constructors

ListEventTypesFilterName'

Fields

fromListEventTypesFilterName :: Text

Bundled Patterns

pattern ListEventTypesFilterName_RESOURCE_TYPE :: ListEventTypesFilterName

pattern ListEventTypesFilterName_SERVICE_NAME :: ListEventTypesFilterName

ListNotificationRulesFilterName

newtype ListNotificationRulesFilterName

Constructors

ListNotificationRulesFilterName'

Fields

fromListNotificationRulesFilterName :: Text

Bundled Patterns

pattern ListNotificationRulesFilterName_CREATED_BY :: ListNotificationRulesFilterName

pattern ListNotificationRulesFilterName_EVENT_TYPE_ID :: ListNotificationRulesFilterName

pattern ListNotificationRulesFilterName_RESOURCE :: ListNotificationRulesFilterName

pattern ListNotificationRulesFilterName_TARGET_ADDRESS :: ListNotificationRulesFilterName

ListTargetsFilterName

newtype ListTargetsFilterName

Constructors

ListTargetsFilterName'

Fields

fromListTargetsFilterName :: Text

Bundled Patterns

pattern ListTargetsFilterName_TARGET_ADDRESS :: ListTargetsFilterName

pattern ListTargetsFilterName_TARGET_STATUS :: ListTargetsFilterName

pattern ListTargetsFilterName_TARGET_TYPE :: ListTargetsFilterName

NotificationRuleStatus

newtype NotificationRuleStatus

Constructors

NotificationRuleStatus'

Fields

fromNotificationRuleStatus :: Text

Bundled Patterns

pattern NotificationRuleStatus_DISABLED :: NotificationRuleStatus

pattern NotificationRuleStatus_ENABLED :: NotificationRuleStatus

TargetStatus

newtype TargetStatus

Constructors

TargetStatus'

Fields

fromTargetStatus :: Text

Bundled Patterns

pattern TargetStatus_ACTIVE :: TargetStatus

pattern TargetStatus_DEACTIVATED :: TargetStatus

pattern TargetStatus_INACTIVE :: TargetStatus

pattern TargetStatus_PENDING :: TargetStatus

pattern TargetStatus_UNREACHABLE :: TargetStatus

EventTypeSummary

data EventTypeSummary

Returns information about an event that has triggered a notification rule.

See: newEventTypeSummary smart constructor.

Constructors

EventTypeSummary'

Fields

resourceType :: Maybe Text

The resource type of the event.

eventTypeName :: Maybe Text

The name of the event.

eventTypeId :: Maybe Text

The system-generated ID of the event.

serviceName :: Maybe Text

The name of the service for which the event applies.

newEventTypeSummary :: EventTypeSummary

Create a value of EventTypeSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceType:EventTypeSummary', eventTypeSummary_resourceType - The resource type of the event.

$sel:eventTypeName:EventTypeSummary', eventTypeSummary_eventTypeName - The name of the event.

$sel:eventTypeId:EventTypeSummary', eventTypeSummary_eventTypeId - The system-generated ID of the event.

$sel:serviceName:EventTypeSummary', eventTypeSummary_serviceName - The name of the service for which the event applies.

eventTypeSummary_resourceType :: Lens' EventTypeSummary (Maybe Text)

The resource type of the event.

eventTypeSummary_eventTypeName :: Lens' EventTypeSummary (Maybe Text)

The name of the event.

eventTypeSummary_eventTypeId :: Lens' EventTypeSummary (Maybe Text)

The system-generated ID of the event.

eventTypeSummary_serviceName :: Lens' EventTypeSummary (Maybe Text)

The name of the service for which the event applies.

ListEventTypesFilter

data ListEventTypesFilter

Information about a filter to apply to the list of returned event types. You can filter by resource type or service name.

See: newListEventTypesFilter smart constructor.

Constructors

ListEventTypesFilter'

Fields

name :: ListEventTypesFilterName

The system-generated name of the filter type you want to filter by.

value :: Text

The name of the resource type (for example, pipeline) or service name (for example, CodePipeline) that you want to filter by.

newListEventTypesFilter

Arguments

:: ListEventTypesFilterName
$sel:name:ListEventTypesFilter'

-> Text
$sel:value:ListEventTypesFilter'

-> ListEventTypesFilter

Create a value of ListEventTypesFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:ListEventTypesFilter', listEventTypesFilter_name - The system-generated name of the filter type you want to filter by.

$sel:value:ListEventTypesFilter', listEventTypesFilter_value - The name of the resource type (for example, pipeline) or service name (for example, CodePipeline) that you want to filter by.

listEventTypesFilter_name :: Lens' ListEventTypesFilter ListEventTypesFilterName

The system-generated name of the filter type you want to filter by.

listEventTypesFilter_value :: Lens' ListEventTypesFilter Text

The name of the resource type (for example, pipeline) or service name (for example, CodePipeline) that you want to filter by.

ListNotificationRulesFilter

data ListNotificationRulesFilter

Information about a filter to apply to the list of returned notification rules. You can filter by event type, owner, resource, or target.

See: newListNotificationRulesFilter smart constructor.

Constructors

ListNotificationRulesFilter'

Fields

name :: ListNotificationRulesFilterName

The name of the attribute you want to use to filter the returned notification rules.

value :: Text

The value of the attribute you want to use to filter the returned notification rules. For example, if you specify filtering by RESOURCE in Name, you might specify the ARN of a pipeline in AWS CodePipeline for the value.

newListNotificationRulesFilter

Arguments

:: ListNotificationRulesFilterName
$sel:name:ListNotificationRulesFilter'

-> Text
$sel:value:ListNotificationRulesFilter'

-> ListNotificationRulesFilter

Create a value of ListNotificationRulesFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:ListNotificationRulesFilter', listNotificationRulesFilter_name - The name of the attribute you want to use to filter the returned notification rules.

$sel:value:ListNotificationRulesFilter', listNotificationRulesFilter_value - The value of the attribute you want to use to filter the returned notification rules. For example, if you specify filtering by RESOURCE in Name, you might specify the ARN of a pipeline in AWS CodePipeline for the value.

listNotificationRulesFilter_name :: Lens' ListNotificationRulesFilter ListNotificationRulesFilterName

The name of the attribute you want to use to filter the returned notification rules.

listNotificationRulesFilter_value :: Lens' ListNotificationRulesFilter Text

The value of the attribute you want to use to filter the returned notification rules. For example, if you specify filtering by RESOURCE in Name, you might specify the ARN of a pipeline in AWS CodePipeline for the value.

ListTargetsFilter

data ListTargetsFilter

Information about a filter to apply to the list of returned targets. You can filter by target type, address, or status. For example, to filter results to notification rules that have active Amazon SNS topics as targets, you could specify a ListTargetsFilter Name as TargetType and a Value of SNS, and a Name of TARGET_STATUS and a Value of ACTIVE.

See: newListTargetsFilter smart constructor.

Constructors

ListTargetsFilter'

Fields

name :: ListTargetsFilterName

The name of the attribute you want to use to filter the returned targets.

value :: Text

The value of the attribute you want to use to filter the returned targets. For example, if you specify SNS for the Target type, you could specify an Amazon Resource Name (ARN) for a topic as the value.

newListTargetsFilter

Arguments

:: ListTargetsFilterName
$sel:name:ListTargetsFilter'

-> Text
$sel:value:ListTargetsFilter'

-> ListTargetsFilter

Create a value of ListTargetsFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:ListTargetsFilter', listTargetsFilter_name - The name of the attribute you want to use to filter the returned targets.

$sel:value:ListTargetsFilter', listTargetsFilter_value - The value of the attribute you want to use to filter the returned targets. For example, if you specify SNS for the Target type, you could specify an Amazon Resource Name (ARN) for a topic as the value.

listTargetsFilter_name :: Lens' ListTargetsFilter ListTargetsFilterName

The name of the attribute you want to use to filter the returned targets.

listTargetsFilter_value :: Lens' ListTargetsFilter Text

The value of the attribute you want to use to filter the returned targets. For example, if you specify SNS for the Target type, you could specify an Amazon Resource Name (ARN) for a topic as the value.

NotificationRuleSummary

data NotificationRuleSummary

Information about a specified notification rule.

See: newNotificationRuleSummary smart constructor.

Constructors

NotificationRuleSummary'

Fields

arn :: Maybe Text

The Amazon Resource Name (ARN) of the notification rule.

id :: Maybe Text

The unique ID of the notification rule.

newNotificationRuleSummary :: NotificationRuleSummary

Create a value of NotificationRuleSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:NotificationRuleSummary', notificationRuleSummary_arn - The Amazon Resource Name (ARN) of the notification rule.

$sel:id:NotificationRuleSummary', notificationRuleSummary_id - The unique ID of the notification rule.

notificationRuleSummary_arn :: Lens' NotificationRuleSummary (Maybe Text)

The Amazon Resource Name (ARN) of the notification rule.

notificationRuleSummary_id :: Lens' NotificationRuleSummary (Maybe Text)

The unique ID of the notification rule.

Target

data Target

Information about the SNS topics associated with a notification rule.

See: newTarget smart constructor.

Constructors

Target'

Fields

targetType :: Maybe Text

The target type. Can be an Amazon SNS topic.

targetAddress :: Maybe (Sensitive Text)

The Amazon Resource Name (ARN) of the SNS topic.

newTarget :: Target

Create a value of Target with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetType:Target', target_targetType - The target type. Can be an Amazon SNS topic.

$sel:targetAddress:Target', target_targetAddress - The Amazon Resource Name (ARN) of the SNS topic.

target_targetType :: Lens' Target (Maybe Text)

The target type. Can be an Amazon SNS topic.

target_targetAddress :: Lens' Target (Maybe Text)

The Amazon Resource Name (ARN) of the SNS topic.

TargetSummary

data TargetSummary

Information about the targets specified for a notification rule.

See: newTargetSummary smart constructor.

Constructors

TargetSummary'

Fields

targetType :: Maybe Text

The type of the target (for example, SNS).

targetAddress :: Maybe (Sensitive Text)

The Amazon Resource Name (ARN) of the SNS topic.

targetStatus :: Maybe TargetStatus

The status of the target.

newTargetSummary :: TargetSummary

Create a value of TargetSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetType:TargetSummary', targetSummary_targetType - The type of the target (for example, SNS).

$sel:targetAddress:TargetSummary', targetSummary_targetAddress - The Amazon Resource Name (ARN) of the SNS topic.

$sel:targetStatus:TargetSummary', targetSummary_targetStatus - The status of the target.

targetSummary_targetType :: Lens' TargetSummary (Maybe Text)

The type of the target (for example, SNS).

targetSummary_targetAddress :: Lens' TargetSummary (Maybe Text)

The Amazon Resource Name (ARN) of the SNS topic.

targetSummary_targetStatus :: Lens' TargetSummary (Maybe TargetStatus)

The status of the target.