Network.AWS.CodeStarNotifications.Types.TargetSummary

Description

Documentation

data TargetSummary

Information about the targets specified for a notification rule.

See: newTargetSummary smart constructor.

Constructors

TargetSummary'

Fields

targetType :: Maybe Text

The type of the target (for example, SNS).

targetAddress :: Maybe (Sensitive Text)

The Amazon Resource Name (ARN) of the SNS topic.

targetStatus :: Maybe TargetStatus

The status of the target.

newTargetSummary :: TargetSummary

Create a value of TargetSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetType:TargetSummary', targetSummary_targetType - The type of the target (for example, SNS).

$sel:targetAddress:TargetSummary', targetSummary_targetAddress - The Amazon Resource Name (ARN) of the SNS topic.

$sel:targetStatus:TargetSummary', targetSummary_targetStatus - The status of the target.

targetSummary_targetType :: Lens' TargetSummary (Maybe Text)

The type of the target (for example, SNS).

targetSummary_targetAddress :: Lens' TargetSummary (Maybe Text)

The Amazon Resource Name (ARN) of the SNS topic.

targetSummary_targetStatus :: Lens' TargetSummary (Maybe TargetStatus)

The status of the target.