Network.AWS.CodeStarNotifications.UpdateNotificationRule

Description

Updates a notification rule for a resource. You can change the events that trigger the notification rule, the status of the rule, and the targets that receive the notifications.

To add or remove tags for a notification rule, you must use TagResource and UntagResource.

Creating a Request

data UpdateNotificationRule

See: newUpdateNotificationRule smart constructor.

Constructors

UpdateNotificationRule'

Fields

status :: Maybe NotificationRuleStatus

The status of the notification rule. Valid statuses include enabled (sending notifications) or disabled (not sending notifications).

eventTypeIds :: Maybe [Text]

A list of event types associated with this notification rule.

detailType :: Maybe DetailType

The level of detail to include in the notifications for this resource. BASIC will include only the contents of the event as it would appear in AWS CloudWatch. FULL will include any supplemental information provided by AWS CodeStar Notifications and/or the service for the resource for which the notification is created.

name :: Maybe (Sensitive Text)

The name of the notification rule.

targets :: Maybe [Target]

The address and type of the targets to receive notifications from this notification rule.

arn :: Text

The Amazon Resource Name (ARN) of the notification rule.

newUpdateNotificationRule

Arguments

:: Text
$sel:arn:UpdateNotificationRule'

-> UpdateNotificationRule

Create a value of UpdateNotificationRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:UpdateNotificationRule', updateNotificationRule_status - The status of the notification rule. Valid statuses include enabled (sending notifications) or disabled (not sending notifications).

$sel:eventTypeIds:UpdateNotificationRule', updateNotificationRule_eventTypeIds - A list of event types associated with this notification rule.

$sel:detailType:UpdateNotificationRule', updateNotificationRule_detailType - The level of detail to include in the notifications for this resource. BASIC will include only the contents of the event as it would appear in AWS CloudWatch. FULL will include any supplemental information provided by AWS CodeStar Notifications and/or the service for the resource for which the notification is created.

$sel:name:UpdateNotificationRule', updateNotificationRule_name - The name of the notification rule.

$sel:targets:UpdateNotificationRule', updateNotificationRule_targets - The address and type of the targets to receive notifications from this notification rule.

$sel:arn:UpdateNotificationRule', updateNotificationRule_arn - The Amazon Resource Name (ARN) of the notification rule.

Request Lenses

updateNotificationRule_status :: Lens' UpdateNotificationRule (Maybe NotificationRuleStatus)

The status of the notification rule. Valid statuses include enabled (sending notifications) or disabled (not sending notifications).

updateNotificationRule_eventTypeIds :: Lens' UpdateNotificationRule (Maybe [Text])

A list of event types associated with this notification rule.

updateNotificationRule_detailType :: Lens' UpdateNotificationRule (Maybe DetailType)

The level of detail to include in the notifications for this resource. BASIC will include only the contents of the event as it would appear in AWS CloudWatch. FULL will include any supplemental information provided by AWS CodeStar Notifications and/or the service for the resource for which the notification is created.

updateNotificationRule_name :: Lens' UpdateNotificationRule (Maybe Text)

The name of the notification rule.

updateNotificationRule_targets :: Lens' UpdateNotificationRule (Maybe [Target])

The address and type of the targets to receive notifications from this notification rule.

updateNotificationRule_arn :: Lens' UpdateNotificationRule Text

The Amazon Resource Name (ARN) of the notification rule.

Destructuring the Response

data UpdateNotificationRuleResponse

See: newUpdateNotificationRuleResponse smart constructor.

Constructors

UpdateNotificationRuleResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateNotificationRuleResponse

Arguments

:: Int
$sel:httpStatus:UpdateNotificationRuleResponse'

-> UpdateNotificationRuleResponse

Create a value of UpdateNotificationRuleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateNotificationRuleResponse', updateNotificationRuleResponse_httpStatus - The response's http status code.

Response Lenses

updateNotificationRuleResponse_httpStatus :: Lens' UpdateNotificationRuleResponse Int

The response's http status code.