Network.AWS.CognitoIdentity.Lens

Description

Operations

DescribeIdentityPool

describeIdentityPool_identityPoolId :: Lens' DescribeIdentityPool Text

An identity pool ID in the format REGION:GUID.

identityPool_allowClassicFlow :: Lens' IdentityPool (Maybe Bool)

Enables or disables the Basic (Classic) authentication flow. For more information, see Identity Pools (Federated Identities) Authentication Flow in the Amazon Cognito Developer Guide.

identityPool_samlProviderARNs :: Lens' IdentityPool (Maybe [Text])

An array of Amazon Resource Names (ARNs) of the SAML provider for your identity pool.

identityPool_identityPoolTags :: Lens' IdentityPool (Maybe (HashMap Text Text))

The tags that are assigned to the identity pool. A tag is a label that you can apply to identity pools to categorize and manage them in different ways, such as by purpose, owner, environment, or other criteria.

identityPool_openIdConnectProviderARNs :: Lens' IdentityPool (Maybe [Text])

The ARNs of the OpenID Connect providers.

identityPool_supportedLoginProviders :: Lens' IdentityPool (Maybe (HashMap Text Text))

Optional key:value pairs mapping provider names to provider app IDs.

identityPool_cognitoIdentityProviders :: Lens' IdentityPool (Maybe [CognitoIdentityProvider])

A list representing an Amazon Cognito user pool and its client ID.

identityPool_developerProviderName :: Lens' IdentityPool (Maybe Text)

The "domain" by which Cognito will refer to your users.

identityPool_identityPoolId :: Lens' IdentityPool Text

An identity pool ID in the format REGION:GUID.

identityPool_identityPoolName :: Lens' IdentityPool Text

A string that you provide.

identityPool_allowUnauthenticatedIdentities :: Lens' IdentityPool Bool

TRUE if the identity pool supports unauthenticated logins.

GetOpenIdTokenForDeveloperIdentity

getOpenIdTokenForDeveloperIdentity_tokenDuration :: Lens' GetOpenIdTokenForDeveloperIdentity (Maybe Natural)

The expiration time of the token, in seconds. You can specify a custom expiration time for the token so that you can cache it. If you don't provide an expiration time, the token is valid for 15 minutes. You can exchange the token with Amazon STS for temporary AWS credentials, which are valid for a maximum of one hour. The maximum token duration you can set is 24 hours. You should take care in setting the expiration time for a token, as there are significant security implications: an attacker could use a leaked token to access your AWS resources for the token's duration.

Please provide for a small grace period, usually no more than 5 minutes, to account for clock skew.

getOpenIdTokenForDeveloperIdentity_principalTags :: Lens' GetOpenIdTokenForDeveloperIdentity (Maybe (HashMap Text Text))

Use this operation to configure attribute mappings for custom providers.

getOpenIdTokenForDeveloperIdentity_identityId :: Lens' GetOpenIdTokenForDeveloperIdentity (Maybe Text)

A unique identifier in the format REGION:GUID.

getOpenIdTokenForDeveloperIdentity_identityPoolId :: Lens' GetOpenIdTokenForDeveloperIdentity Text

An identity pool ID in the format REGION:GUID.

getOpenIdTokenForDeveloperIdentity_logins :: Lens' GetOpenIdTokenForDeveloperIdentity (HashMap Text Text)

A set of optional name-value pairs that map provider names to provider tokens. Each name-value pair represents a user from a public provider or developer provider. If the user is from a developer provider, the name-value pair will follow the syntax "developer_provider_name": "developer_user_identifier". The developer provider is the "domain" by which Cognito will refer to your users; you provided this domain while creating/updating the identity pool. The developer user identifier is an identifier from your backend that uniquely identifies a user. When you create an identity pool, you can specify the supported logins.

getOpenIdTokenForDeveloperIdentityResponse_identityId :: Lens' GetOpenIdTokenForDeveloperIdentityResponse (Maybe Text)

A unique identifier in the format REGION:GUID.

getOpenIdTokenForDeveloperIdentityResponse_token :: Lens' GetOpenIdTokenForDeveloperIdentityResponse (Maybe Text)

An OpenID token.

getOpenIdTokenForDeveloperIdentityResponse_httpStatus :: Lens' GetOpenIdTokenForDeveloperIdentityResponse Int

The response's http status code.

GetOpenIdToken

getOpenIdToken_logins :: Lens' GetOpenIdToken (Maybe (HashMap Text Text))

A set of optional name-value pairs that map provider names to provider tokens. When using graph.facebook.com and www.amazon.com, supply the access_token returned from the provider's authflow. For accounts.google.com, an Amazon Cognito user pool provider, or any other OpenID Connect provider, always include the id_token.

getOpenIdToken_identityId :: Lens' GetOpenIdToken Text

A unique identifier in the format REGION:GUID.

getOpenIdTokenResponse_identityId :: Lens' GetOpenIdTokenResponse (Maybe Text)

A unique identifier in the format REGION:GUID. Note that the IdentityId returned may not match the one passed on input.

getOpenIdTokenResponse_token :: Lens' GetOpenIdTokenResponse (Maybe Text)

An OpenID token, valid for 10 minutes.

getOpenIdTokenResponse_httpStatus :: Lens' GetOpenIdTokenResponse Int

The response's http status code.

DeleteIdentities

deleteIdentities_identityIdsToDelete :: Lens' DeleteIdentities (NonEmpty Text)

A list of 1-60 identities that you want to delete.

deleteIdentitiesResponse_unprocessedIdentityIds :: Lens' DeleteIdentitiesResponse (Maybe [UnprocessedIdentityId])

An array of UnprocessedIdentityId objects, each of which contains an ErrorCode and IdentityId.

deleteIdentitiesResponse_httpStatus :: Lens' DeleteIdentitiesResponse Int

The response's http status code.

GetPrincipalTagAttributeMap

getPrincipalTagAttributeMap_identityPoolId :: Lens' GetPrincipalTagAttributeMap Text

You can use this operation to get the ID of the Identity Pool you setup attribute mappings for.

getPrincipalTagAttributeMap_identityProviderName :: Lens' GetPrincipalTagAttributeMap Text

You can use this operation to get the provider name.

getPrincipalTagAttributeMapResponse_identityPoolId :: Lens' GetPrincipalTagAttributeMapResponse (Maybe Text)

You can use this operation to get the ID of the Identity Pool you setup attribute mappings for.

getPrincipalTagAttributeMapResponse_identityProviderName :: Lens' GetPrincipalTagAttributeMapResponse (Maybe Text)

You can use this operation to get the provider name.

getPrincipalTagAttributeMapResponse_principalTags :: Lens' GetPrincipalTagAttributeMapResponse (Maybe (HashMap Text Text))

You can use this operation to add principal tags. The PrincipalTagsoperation enables you to reference user attributes in your IAM permissions policy.

getPrincipalTagAttributeMapResponse_useDefaults :: Lens' GetPrincipalTagAttributeMapResponse (Maybe Bool)

You can use this operation to list

getPrincipalTagAttributeMapResponse_httpStatus :: Lens' GetPrincipalTagAttributeMapResponse Int

The response's http status code.

CreateIdentityPool

createIdentityPool_allowClassicFlow :: Lens' CreateIdentityPool (Maybe Bool)

Enables or disables the Basic (Classic) authentication flow. For more information, see Identity Pools (Federated Identities) Authentication Flow in the Amazon Cognito Developer Guide.

createIdentityPool_samlProviderARNs :: Lens' CreateIdentityPool (Maybe [Text])

An array of Amazon Resource Names (ARNs) of the SAML provider for your identity pool.

createIdentityPool_identityPoolTags :: Lens' CreateIdentityPool (Maybe (HashMap Text Text))

Tags to assign to the identity pool. A tag is a label that you can apply to identity pools to categorize and manage them in different ways, such as by purpose, owner, environment, or other criteria.

createIdentityPool_openIdConnectProviderARNs :: Lens' CreateIdentityPool (Maybe [Text])

The Amazon Resource Names (ARN) of the OpenID Connect providers.

createIdentityPool_supportedLoginProviders :: Lens' CreateIdentityPool (Maybe (HashMap Text Text))

Optional key:value pairs mapping provider names to provider app IDs.

createIdentityPool_cognitoIdentityProviders :: Lens' CreateIdentityPool (Maybe [CognitoIdentityProvider])

An array of Amazon Cognito user pools and their client IDs.

createIdentityPool_developerProviderName :: Lens' CreateIdentityPool (Maybe Text)

The "domain" by which Cognito will refer to your users. This name acts as a placeholder that allows your backend and the Cognito service to communicate about the developer provider. For the DeveloperProviderName, you can use letters as well as period (.), underscore (_), and dash (-).

Once you have set a developer provider name, you cannot change it. Please take care in setting this parameter.

createIdentityPool_identityPoolName :: Lens' CreateIdentityPool Text

A string that you provide.

createIdentityPool_allowUnauthenticatedIdentities :: Lens' CreateIdentityPool Bool

TRUE if the identity pool supports unauthenticated logins.

identityPool_allowClassicFlow :: Lens' IdentityPool (Maybe Bool)

Enables or disables the Basic (Classic) authentication flow. For more information, see Identity Pools (Federated Identities) Authentication Flow in the Amazon Cognito Developer Guide.

identityPool_samlProviderARNs :: Lens' IdentityPool (Maybe [Text])

An array of Amazon Resource Names (ARNs) of the SAML provider for your identity pool.

identityPool_identityPoolTags :: Lens' IdentityPool (Maybe (HashMap Text Text))

The tags that are assigned to the identity pool. A tag is a label that you can apply to identity pools to categorize and manage them in different ways, such as by purpose, owner, environment, or other criteria.

identityPool_openIdConnectProviderARNs :: Lens' IdentityPool (Maybe [Text])

The ARNs of the OpenID Connect providers.

identityPool_supportedLoginProviders :: Lens' IdentityPool (Maybe (HashMap Text Text))

Optional key:value pairs mapping provider names to provider app IDs.

identityPool_cognitoIdentityProviders :: Lens' IdentityPool (Maybe [CognitoIdentityProvider])

A list representing an Amazon Cognito user pool and its client ID.

identityPool_developerProviderName :: Lens' IdentityPool (Maybe Text)

The "domain" by which Cognito will refer to your users.

identityPool_identityPoolId :: Lens' IdentityPool Text

An identity pool ID in the format REGION:GUID.

identityPool_identityPoolName :: Lens' IdentityPool Text

A string that you provide.

identityPool_allowUnauthenticatedIdentities :: Lens' IdentityPool Bool

TRUE if the identity pool supports unauthenticated logins.

MergeDeveloperIdentities

mergeDeveloperIdentities_sourceUserIdentifier :: Lens' MergeDeveloperIdentities Text

User identifier for the source user. The value should be a DeveloperUserIdentifier.

mergeDeveloperIdentities_destinationUserIdentifier :: Lens' MergeDeveloperIdentities Text

User identifier for the destination user. The value should be a DeveloperUserIdentifier.

mergeDeveloperIdentities_developerProviderName :: Lens' MergeDeveloperIdentities Text

The "domain" by which Cognito will refer to your users. This is a (pseudo) domain name that you provide while creating an identity pool. This name acts as a placeholder that allows your backend and the Cognito service to communicate about the developer provider. For the DeveloperProviderName, you can use letters as well as period (.), underscore (_), and dash (-).

mergeDeveloperIdentities_identityPoolId :: Lens' MergeDeveloperIdentities Text

An identity pool ID in the format REGION:GUID.

mergeDeveloperIdentitiesResponse_identityId :: Lens' MergeDeveloperIdentitiesResponse (Maybe Text)

A unique identifier in the format REGION:GUID.

mergeDeveloperIdentitiesResponse_httpStatus :: Lens' MergeDeveloperIdentitiesResponse Int

The response's http status code.

UpdateIdentityPool

updateIdentityPool_allowClassicFlow :: Lens' UpdateIdentityPool (Maybe Bool)

Enables or disables the Basic (Classic) authentication flow. For more information, see Identity Pools (Federated Identities) Authentication Flow in the Amazon Cognito Developer Guide.

updateIdentityPool_samlProviderARNs :: Lens' UpdateIdentityPool (Maybe [Text])

An array of Amazon Resource Names (ARNs) of the SAML provider for your identity pool.

updateIdentityPool_identityPoolTags :: Lens' UpdateIdentityPool (Maybe (HashMap Text Text))

The tags that are assigned to the identity pool. A tag is a label that you can apply to identity pools to categorize and manage them in different ways, such as by purpose, owner, environment, or other criteria.

updateIdentityPool_openIdConnectProviderARNs :: Lens' UpdateIdentityPool (Maybe [Text])

The ARNs of the OpenID Connect providers.

updateIdentityPool_supportedLoginProviders :: Lens' UpdateIdentityPool (Maybe (HashMap Text Text))

Optional key:value pairs mapping provider names to provider app IDs.

updateIdentityPool_cognitoIdentityProviders :: Lens' UpdateIdentityPool (Maybe [CognitoIdentityProvider])

A list representing an Amazon Cognito user pool and its client ID.

updateIdentityPool_developerProviderName :: Lens' UpdateIdentityPool (Maybe Text)

The "domain" by which Cognito will refer to your users.

updateIdentityPool_identityPoolId :: Lens' UpdateIdentityPool Text

An identity pool ID in the format REGION:GUID.

updateIdentityPool_identityPoolName :: Lens' UpdateIdentityPool Text

A string that you provide.

updateIdentityPool_allowUnauthenticatedIdentities :: Lens' UpdateIdentityPool Bool

TRUE if the identity pool supports unauthenticated logins.

identityPool_allowClassicFlow :: Lens' IdentityPool (Maybe Bool)

Enables or disables the Basic (Classic) authentication flow. For more information, see Identity Pools (Federated Identities) Authentication Flow in the Amazon Cognito Developer Guide.

identityPool_samlProviderARNs :: Lens' IdentityPool (Maybe [Text])

An array of Amazon Resource Names (ARNs) of the SAML provider for your identity pool.

identityPool_identityPoolTags :: Lens' IdentityPool (Maybe (HashMap Text Text))

The tags that are assigned to the identity pool. A tag is a label that you can apply to identity pools to categorize and manage them in different ways, such as by purpose, owner, environment, or other criteria.

identityPool_openIdConnectProviderARNs :: Lens' IdentityPool (Maybe [Text])

The ARNs of the OpenID Connect providers.

identityPool_supportedLoginProviders :: Lens' IdentityPool (Maybe (HashMap Text Text))

Optional key:value pairs mapping provider names to provider app IDs.

identityPool_cognitoIdentityProviders :: Lens' IdentityPool (Maybe [CognitoIdentityProvider])

A list representing an Amazon Cognito user pool and its client ID.

identityPool_developerProviderName :: Lens' IdentityPool (Maybe Text)

The "domain" by which Cognito will refer to your users.

identityPool_identityPoolId :: Lens' IdentityPool Text

An identity pool ID in the format REGION:GUID.

identityPool_identityPoolName :: Lens' IdentityPool Text

A string that you provide.

identityPool_allowUnauthenticatedIdentities :: Lens' IdentityPool Bool

TRUE if the identity pool supports unauthenticated logins.

GetIdentityPoolRoles

getIdentityPoolRoles_identityPoolId :: Lens' GetIdentityPoolRoles Text

An identity pool ID in the format REGION:GUID.

getIdentityPoolRolesResponse_identityPoolId :: Lens' GetIdentityPoolRolesResponse (Maybe Text)

An identity pool ID in the format REGION:GUID.

getIdentityPoolRolesResponse_roles :: Lens' GetIdentityPoolRolesResponse (Maybe (HashMap Text Text))

The map of roles associated with this pool. Currently only authenticated and unauthenticated roles are supported.

getIdentityPoolRolesResponse_roleMappings :: Lens' GetIdentityPoolRolesResponse (Maybe (HashMap Text RoleMapping))

How users for a specific identity provider are to mapped to roles. This is a String-to-RoleMapping object map. The string identifies the identity provider, for example, "graph.facebook.com" or "cognito-idp.us-east-1.amazonaws.com/us-east-1_abcdefghi:app_client_id".

getIdentityPoolRolesResponse_httpStatus :: Lens' GetIdentityPoolRolesResponse Int

The response's http status code.

UntagResource

untagResource_resourceArn :: Lens' UntagResource Text

The Amazon Resource Name (ARN) of the identity pool.

untagResource_tagKeys :: Lens' UntagResource [Text]

The keys of the tags to remove from the user pool.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

DeleteIdentityPool

deleteIdentityPool_identityPoolId :: Lens' DeleteIdentityPool Text

An identity pool ID in the format REGION:GUID.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The Amazon Resource Name (ARN) of the identity pool.

tagResource_tags :: Lens' TagResource (HashMap Text Text)

The tags to assign to the identity pool.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

UnlinkIdentity

unlinkIdentity_identityId :: Lens' UnlinkIdentity Text

A unique identifier in the format REGION:GUID.

unlinkIdentity_logins :: Lens' UnlinkIdentity (HashMap Text Text)

A set of optional name-value pairs that map provider names to provider tokens.

unlinkIdentity_loginsToRemove :: Lens' UnlinkIdentity [Text]

Provider names to unlink from this identity.

LookupDeveloperIdentity

lookupDeveloperIdentity_nextToken :: Lens' LookupDeveloperIdentity (Maybe Text)

A pagination token. The first call you make will have NextToken set to null. After that the service will return NextToken values as needed. For example, let's say you make a request with MaxResults set to 10, and there are 20 matches in the database. The service will return a pagination token as a part of the response. This token can be used to call the API again and get results starting from the 11th match.

lookupDeveloperIdentity_maxResults :: Lens' LookupDeveloperIdentity (Maybe Natural)

The maximum number of identities to return.

lookupDeveloperIdentity_developerUserIdentifier :: Lens' LookupDeveloperIdentity (Maybe Text)

A unique ID used by your backend authentication process to identify a user. Typically, a developer identity provider would issue many developer user identifiers, in keeping with the number of users.

lookupDeveloperIdentity_identityId :: Lens' LookupDeveloperIdentity (Maybe Text)

A unique identifier in the format REGION:GUID.

lookupDeveloperIdentity_identityPoolId :: Lens' LookupDeveloperIdentity Text

An identity pool ID in the format REGION:GUID.

lookupDeveloperIdentityResponse_nextToken :: Lens' LookupDeveloperIdentityResponse (Maybe Text)

A pagination token. The first call you make will have NextToken set to null. After that the service will return NextToken values as needed. For example, let's say you make a request with MaxResults set to 10, and there are 20 matches in the database. The service will return a pagination token as a part of the response. This token can be used to call the API again and get results starting from the 11th match.

lookupDeveloperIdentityResponse_developerUserIdentifierList :: Lens' LookupDeveloperIdentityResponse (Maybe [Text])

This is the list of developer user identifiers associated with an identity ID. Cognito supports the association of multiple developer user identifiers with an identity ID.

lookupDeveloperIdentityResponse_identityId :: Lens' LookupDeveloperIdentityResponse (Maybe Text)

A unique identifier in the format REGION:GUID.

lookupDeveloperIdentityResponse_httpStatus :: Lens' LookupDeveloperIdentityResponse Int

The response's http status code.

SetIdentityPoolRoles

setIdentityPoolRoles_roleMappings :: Lens' SetIdentityPoolRoles (Maybe (HashMap Text RoleMapping))

How users for a specific identity provider are to mapped to roles. This is a string to RoleMapping object map. The string identifies the identity provider, for example, "graph.facebook.com" or "cognito-idp.us-east-1.amazonaws.com/us-east-1_abcdefghi:app_client_id".

Up to 25 rules can be specified per identity provider.

setIdentityPoolRoles_identityPoolId :: Lens' SetIdentityPoolRoles Text

An identity pool ID in the format REGION:GUID.

setIdentityPoolRoles_roles :: Lens' SetIdentityPoolRoles (HashMap Text Text)

The map of roles associated with this pool. For a given role, the key will be either "authenticated" or "unauthenticated" and the value will be the Role ARN.

ListIdentityPools

listIdentityPools_nextToken :: Lens' ListIdentityPools (Maybe Text)

A pagination token.

listIdentityPools_maxResults :: Lens' ListIdentityPools Natural

The maximum number of identities to return.

listIdentityPoolsResponse_nextToken :: Lens' ListIdentityPoolsResponse (Maybe Text)

A pagination token.

listIdentityPoolsResponse_identityPools :: Lens' ListIdentityPoolsResponse (Maybe [IdentityPoolShortDescription])

The identity pools returned by the ListIdentityPools action.

listIdentityPoolsResponse_httpStatus :: Lens' ListIdentityPoolsResponse Int

The response's http status code.

GetId

getId_accountId :: Lens' GetId (Maybe Text)

A standard AWS account ID (9+ digits).

getId_logins :: Lens' GetId (Maybe (HashMap Text Text))

A set of optional name-value pairs that map provider names to provider tokens. The available provider names for Logins are as follows:

Facebook: graph.facebook.com

Amazon Cognito user pool: cognito-idp.<region>.amazonaws.com/<YOUR_USER_POOL_ID>, for example, cognito-idp.us-east-1.amazonaws.com/us-east-1_123456789.

Google: accounts.google.com

Amazon: www.amazon.com

Twitter: api.twitter.com

Digits: www.digits.com

getId_identityPoolId :: Lens' GetId Text

An identity pool ID in the format REGION:GUID.

getIdResponse_identityId :: Lens' GetIdResponse (Maybe Text)

A unique identifier in the format REGION:GUID.

getIdResponse_httpStatus :: Lens' GetIdResponse Int

The response's http status code.

DescribeIdentity

describeIdentity_identityId :: Lens' DescribeIdentity Text

A unique identifier in the format REGION:GUID.

identityDescription_lastModifiedDate :: Lens' IdentityDescription (Maybe UTCTime)

Date on which the identity was last modified.

identityDescription_creationDate :: Lens' IdentityDescription (Maybe UTCTime)

Date on which the identity was created.

identityDescription_identityId :: Lens' IdentityDescription (Maybe Text)

A unique identifier in the format REGION:GUID.

identityDescription_logins :: Lens' IdentityDescription (Maybe [Text])

The provider names.

GetCredentialsForIdentity

getCredentialsForIdentity_logins :: Lens' GetCredentialsForIdentity (Maybe (HashMap Text Text))

A set of optional name-value pairs that map provider names to provider tokens. The name-value pair will follow the syntax "provider_name": "provider_user_identifier".

Logins should not be specified when trying to get credentials for an unauthenticated identity.

The Logins parameter is required when using identities associated with external identity providers such as Facebook. For examples of Logins maps, see the code examples in the External Identity Providers section of the Amazon Cognito Developer Guide.

getCredentialsForIdentity_customRoleArn :: Lens' GetCredentialsForIdentity (Maybe Text)

The Amazon Resource Name (ARN) of the role to be assumed when multiple roles were received in the token from the identity provider. For example, a SAML-based identity provider. This parameter is optional for identity providers that do not support role customization.

getCredentialsForIdentity_identityId :: Lens' GetCredentialsForIdentity Text

A unique identifier in the format REGION:GUID.

getCredentialsForIdentityResponse_identityId :: Lens' GetCredentialsForIdentityResponse (Maybe Text)

A unique identifier in the format REGION:GUID.

getCredentialsForIdentityResponse_credentials :: Lens' GetCredentialsForIdentityResponse (Maybe Credentials)

Credentials for the provided identity ID.

getCredentialsForIdentityResponse_httpStatus :: Lens' GetCredentialsForIdentityResponse Int

The response's http status code.

UnlinkDeveloperIdentity

unlinkDeveloperIdentity_identityId :: Lens' UnlinkDeveloperIdentity Text

A unique identifier in the format REGION:GUID.

unlinkDeveloperIdentity_identityPoolId :: Lens' UnlinkDeveloperIdentity Text

An identity pool ID in the format REGION:GUID.

unlinkDeveloperIdentity_developerProviderName :: Lens' UnlinkDeveloperIdentity Text

The "domain" by which Cognito will refer to your users.

unlinkDeveloperIdentity_developerUserIdentifier :: Lens' UnlinkDeveloperIdentity Text

A unique ID used by your backend authentication process to identify a user.

ListIdentities

listIdentities_nextToken :: Lens' ListIdentities (Maybe Text)

A pagination token.

listIdentities_hideDisabled :: Lens' ListIdentities (Maybe Bool)

An optional boolean parameter that allows you to hide disabled identities. If omitted, the ListIdentities API will include disabled identities in the response.

listIdentities_identityPoolId :: Lens' ListIdentities Text

An identity pool ID in the format REGION:GUID.

listIdentities_maxResults :: Lens' ListIdentities Natural

The maximum number of identities to return.

listIdentitiesResponse_identityPoolId :: Lens' ListIdentitiesResponse (Maybe Text)

An identity pool ID in the format REGION:GUID.

listIdentitiesResponse_nextToken :: Lens' ListIdentitiesResponse (Maybe Text)

A pagination token.

listIdentitiesResponse_identities :: Lens' ListIdentitiesResponse (Maybe [IdentityDescription])

An object containing a set of identities and associated mappings.

listIdentitiesResponse_httpStatus :: Lens' ListIdentitiesResponse Int

The response's http status code.

SetPrincipalTagAttributeMap

setPrincipalTagAttributeMap_principalTags :: Lens' SetPrincipalTagAttributeMap (Maybe (HashMap Text Text))

You can use this operation to add principal tags.

setPrincipalTagAttributeMap_useDefaults :: Lens' SetPrincipalTagAttributeMap (Maybe Bool)

You can use this operation to use default (username and clientID) attribute mappings.

setPrincipalTagAttributeMap_identityPoolId :: Lens' SetPrincipalTagAttributeMap Text

The ID of the Identity Pool you want to set attribute mappings for.

setPrincipalTagAttributeMap_identityProviderName :: Lens' SetPrincipalTagAttributeMap Text

The provider name you want to use for attribute mappings.

setPrincipalTagAttributeMapResponse_identityPoolId :: Lens' SetPrincipalTagAttributeMapResponse (Maybe Text)

The ID of the Identity Pool you want to set attribute mappings for.

setPrincipalTagAttributeMapResponse_identityProviderName :: Lens' SetPrincipalTagAttributeMapResponse (Maybe Text)

The provider name you want to use for attribute mappings.

setPrincipalTagAttributeMapResponse_principalTags :: Lens' SetPrincipalTagAttributeMapResponse (Maybe (HashMap Text Text))

You can use this operation to add principal tags. The PrincipalTagsoperation enables you to reference user attributes in your IAM permissions policy.

setPrincipalTagAttributeMapResponse_useDefaults :: Lens' SetPrincipalTagAttributeMapResponse (Maybe Bool)

You can use this operation to select default (username and clientID) attribute mappings.

setPrincipalTagAttributeMapResponse_httpStatus :: Lens' SetPrincipalTagAttributeMapResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The Amazon Resource Name (ARN) of the identity pool that the tags are assigned to.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe (HashMap Text Text))

The tags that are assigned to the identity pool.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

Types

CognitoIdentityProvider

cognitoIdentityProvider_clientId :: Lens' CognitoIdentityProvider (Maybe Text)

The client ID for the Amazon Cognito user pool.

cognitoIdentityProvider_providerName :: Lens' CognitoIdentityProvider (Maybe Text)

The provider name for an Amazon Cognito user pool. For example, cognito-idp.us-east-1.amazonaws.com/us-east-1_123456789.

cognitoIdentityProvider_serverSideTokenCheck :: Lens' CognitoIdentityProvider (Maybe Bool)

TRUE if server-side token validation is enabled for the identity provider’s token.

Once you set ServerSideTokenCheck to TRUE for an identity pool, that identity pool will check with the integrated user pools to make sure that the user has not been globally signed out or deleted before the identity pool provides an OIDC token or AWS credentials for the user.

If the user is signed out or deleted, the identity pool will return a 400 Not Authorized error.

Credentials

credentials_expiration :: Lens' Credentials (Maybe UTCTime)

The date at which these credentials will expire.

credentials_secretKey :: Lens' Credentials (Maybe Text)

The Secret Access Key portion of the credentials

credentials_accessKeyId :: Lens' Credentials (Maybe Text)

The Access Key portion of the credentials.

credentials_sessionToken :: Lens' Credentials (Maybe Text)

The Session Token portion of the credentials

IdentityDescription

identityDescription_lastModifiedDate :: Lens' IdentityDescription (Maybe UTCTime)

Date on which the identity was last modified.

identityDescription_creationDate :: Lens' IdentityDescription (Maybe UTCTime)

Date on which the identity was created.

identityDescription_identityId :: Lens' IdentityDescription (Maybe Text)

A unique identifier in the format REGION:GUID.

identityDescription_logins :: Lens' IdentityDescription (Maybe [Text])

The provider names.

IdentityPool

identityPool_allowClassicFlow :: Lens' IdentityPool (Maybe Bool)

Enables or disables the Basic (Classic) authentication flow. For more information, see Identity Pools (Federated Identities) Authentication Flow in the Amazon Cognito Developer Guide.

identityPool_samlProviderARNs :: Lens' IdentityPool (Maybe [Text])

An array of Amazon Resource Names (ARNs) of the SAML provider for your identity pool.

identityPool_identityPoolTags :: Lens' IdentityPool (Maybe (HashMap Text Text))

The tags that are assigned to the identity pool. A tag is a label that you can apply to identity pools to categorize and manage them in different ways, such as by purpose, owner, environment, or other criteria.

identityPool_openIdConnectProviderARNs :: Lens' IdentityPool (Maybe [Text])

The ARNs of the OpenID Connect providers.

identityPool_supportedLoginProviders :: Lens' IdentityPool (Maybe (HashMap Text Text))

Optional key:value pairs mapping provider names to provider app IDs.

identityPool_cognitoIdentityProviders :: Lens' IdentityPool (Maybe [CognitoIdentityProvider])

A list representing an Amazon Cognito user pool and its client ID.

identityPool_developerProviderName :: Lens' IdentityPool (Maybe Text)

The "domain" by which Cognito will refer to your users.

identityPool_identityPoolId :: Lens' IdentityPool Text

An identity pool ID in the format REGION:GUID.

identityPool_identityPoolName :: Lens' IdentityPool Text

A string that you provide.

identityPool_allowUnauthenticatedIdentities :: Lens' IdentityPool Bool

TRUE if the identity pool supports unauthenticated logins.

IdentityPoolShortDescription

identityPoolShortDescription_identityPoolId :: Lens' IdentityPoolShortDescription (Maybe Text)

An identity pool ID in the format REGION:GUID.

identityPoolShortDescription_identityPoolName :: Lens' IdentityPoolShortDescription (Maybe Text)

A string that you provide.

MappingRule

mappingRule_claim :: Lens' MappingRule Text

The claim name that must be present in the token, for example, "isAdmin" or "paid".

mappingRule_matchType :: Lens' MappingRule MappingRuleMatchType

The match condition that specifies how closely the claim value in the IdP token must match Value.

mappingRule_value :: Lens' MappingRule Text

A brief string that the claim must match, for example, "paid" or "yes".

mappingRule_roleARN :: Lens' MappingRule Text

The role ARN.

RoleMapping

roleMapping_ambiguousRoleResolution :: Lens' RoleMapping (Maybe AmbiguousRoleResolutionType)

If you specify Token or Rules as the Type, AmbiguousRoleResolution is required.

Specifies the action to be taken if either no rules match the claim value for the Rules type, or there is no cognito:preferred_role claim and there are multiple cognito:roles matches for the Token type.

roleMapping_rulesConfiguration :: Lens' RoleMapping (Maybe RulesConfigurationType)

The rules to be used for mapping users to roles.

If you specify Rules as the role mapping type, RulesConfiguration is required.

roleMapping_type :: Lens' RoleMapping RoleMappingType

The role mapping type. Token will use cognito:roles and cognito:preferred_role claims from the Cognito identity provider token to map groups to roles. Rules will attempt to match claims from the token to map to a role.

RulesConfigurationType

rulesConfigurationType_rules :: Lens' RulesConfigurationType (NonEmpty MappingRule)

An array of rules. You can specify up to 25 rules per identity provider.

Rules are evaluated in order. The first one to match specifies the role.

UnprocessedIdentityId

unprocessedIdentityId_identityId :: Lens' UnprocessedIdentityId (Maybe Text)

A unique identifier in the format REGION:GUID.

unprocessedIdentityId_errorCode :: Lens' UnprocessedIdentityId (Maybe CognitoErrorCode)

The error code indicating the type of error that occurred.