Network.AWS.CognitoIdentity.MergeDeveloperIdentities

Description

Merges two users having different IdentityIds, existing in the same identity pool, and identified by the same developer provider. You can use this action to request that discrete users be merged and identified as a single user in the Cognito environment. Cognito associates the given source user (SourceUserIdentifier) with the IdentityId of the DestinationUserIdentifier. Only developer-authenticated users can be merged. If the users to be merged are associated with the same public provider, but as two different users, an exception will be thrown.

The number of linked logins is limited to 20. So, the number of linked logins for the source user, SourceUserIdentifier, and the destination user, DestinationUserIdentifier, together should not be larger than 20. Otherwise, an exception will be thrown.

You must use AWS Developer credentials to call this API.

Creating a Request

data MergeDeveloperIdentities

Input to the MergeDeveloperIdentities action.

See: newMergeDeveloperIdentities smart constructor.

Constructors

MergeDeveloperIdentities'

Fields

sourceUserIdentifier :: Text

User identifier for the source user. The value should be a DeveloperUserIdentifier.

destinationUserIdentifier :: Text

User identifier for the destination user. The value should be a DeveloperUserIdentifier.

developerProviderName :: Text

The "domain" by which Cognito will refer to your users. This is a (pseudo) domain name that you provide while creating an identity pool. This name acts as a placeholder that allows your backend and the Cognito service to communicate about the developer provider. For the DeveloperProviderName, you can use letters as well as period (.), underscore (_), and dash (-).

identityPoolId :: Text

An identity pool ID in the format REGION:GUID.

newMergeDeveloperIdentities

Arguments

:: Text
$sel:sourceUserIdentifier:MergeDeveloperIdentities'

-> Text
$sel:destinationUserIdentifier:MergeDeveloperIdentities'

-> Text
$sel:developerProviderName:MergeDeveloperIdentities'

-> Text
$sel:identityPoolId:MergeDeveloperIdentities'

-> MergeDeveloperIdentities

Create a value of MergeDeveloperIdentities with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceUserIdentifier:MergeDeveloperIdentities', mergeDeveloperIdentities_sourceUserIdentifier - User identifier for the source user. The value should be a DeveloperUserIdentifier.

$sel:destinationUserIdentifier:MergeDeveloperIdentities', mergeDeveloperIdentities_destinationUserIdentifier - User identifier for the destination user. The value should be a DeveloperUserIdentifier.

$sel:developerProviderName:MergeDeveloperIdentities', mergeDeveloperIdentities_developerProviderName - The "domain" by which Cognito will refer to your users. This is a (pseudo) domain name that you provide while creating an identity pool. This name acts as a placeholder that allows your backend and the Cognito service to communicate about the developer provider. For the DeveloperProviderName, you can use letters as well as period (.), underscore (_), and dash (-).

$sel:identityPoolId:MergeDeveloperIdentities', mergeDeveloperIdentities_identityPoolId - An identity pool ID in the format REGION:GUID.

Request Lenses

mergeDeveloperIdentities_sourceUserIdentifier :: Lens' MergeDeveloperIdentities Text

User identifier for the source user. The value should be a DeveloperUserIdentifier.

mergeDeveloperIdentities_destinationUserIdentifier :: Lens' MergeDeveloperIdentities Text

User identifier for the destination user. The value should be a DeveloperUserIdentifier.

mergeDeveloperIdentities_developerProviderName :: Lens' MergeDeveloperIdentities Text

The "domain" by which Cognito will refer to your users. This is a (pseudo) domain name that you provide while creating an identity pool. This name acts as a placeholder that allows your backend and the Cognito service to communicate about the developer provider. For the DeveloperProviderName, you can use letters as well as period (.), underscore (_), and dash (-).

mergeDeveloperIdentities_identityPoolId :: Lens' MergeDeveloperIdentities Text

An identity pool ID in the format REGION:GUID.

Destructuring the Response

data MergeDeveloperIdentitiesResponse

Returned in response to a successful MergeDeveloperIdentities action.

See: newMergeDeveloperIdentitiesResponse smart constructor.

Constructors

MergeDeveloperIdentitiesResponse'

Fields

identityId :: Maybe Text

A unique identifier in the format REGION:GUID.

httpStatus :: Int

The response's http status code.

newMergeDeveloperIdentitiesResponse

Arguments

:: Int
$sel:httpStatus:MergeDeveloperIdentitiesResponse'

-> MergeDeveloperIdentitiesResponse

Create a value of MergeDeveloperIdentitiesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identityId:MergeDeveloperIdentitiesResponse', mergeDeveloperIdentitiesResponse_identityId - A unique identifier in the format REGION:GUID.

$sel:httpStatus:MergeDeveloperIdentitiesResponse', mergeDeveloperIdentitiesResponse_httpStatus - The response's http status code.

Response Lenses

mergeDeveloperIdentitiesResponse_identityId :: Lens' MergeDeveloperIdentitiesResponse (Maybe Text)

A unique identifier in the format REGION:GUID.

mergeDeveloperIdentitiesResponse_httpStatus :: Lens' MergeDeveloperIdentitiesResponse Int

The response's http status code.