Network.AWS.CognitoIdentity.SetIdentityPoolRoles

Description

Sets the roles for an identity pool. These roles are used when making calls to GetCredentialsForIdentity action.

You must use AWS Developer credentials to call this API.

Creating a Request

data SetIdentityPoolRoles

Input to the SetIdentityPoolRoles action.

See: newSetIdentityPoolRoles smart constructor.

Constructors

SetIdentityPoolRoles'

Fields

roleMappings :: Maybe (HashMap Text RoleMapping)

How users for a specific identity provider are to mapped to roles. This is a string to RoleMapping object map. The string identifies the identity provider, for example, "graph.facebook.com" or "cognito-idp.us-east-1.amazonaws.com/us-east-1_abcdefghi:app_client_id".

Up to 25 rules can be specified per identity provider.

identityPoolId :: Text

An identity pool ID in the format REGION:GUID.

roles :: HashMap Text Text

The map of roles associated with this pool. For a given role, the key will be either "authenticated" or "unauthenticated" and the value will be the Role ARN.

newSetIdentityPoolRoles

Arguments

:: Text
$sel:identityPoolId:SetIdentityPoolRoles'

-> SetIdentityPoolRoles

Create a value of SetIdentityPoolRoles with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:roleMappings:SetIdentityPoolRoles', setIdentityPoolRoles_roleMappings - How users for a specific identity provider are to mapped to roles. This is a string to RoleMapping object map. The string identifies the identity provider, for example, "graph.facebook.com" or "cognito-idp.us-east-1.amazonaws.com/us-east-1_abcdefghi:app_client_id".

Up to 25 rules can be specified per identity provider.

$sel:identityPoolId:SetIdentityPoolRoles', setIdentityPoolRoles_identityPoolId - An identity pool ID in the format REGION:GUID.

$sel:roles:SetIdentityPoolRoles', setIdentityPoolRoles_roles - The map of roles associated with this pool. For a given role, the key will be either "authenticated" or "unauthenticated" and the value will be the Role ARN.

Request Lenses

setIdentityPoolRoles_roleMappings :: Lens' SetIdentityPoolRoles (Maybe (HashMap Text RoleMapping))

How users for a specific identity provider are to mapped to roles. This is a string to RoleMapping object map. The string identifies the identity provider, for example, "graph.facebook.com" or "cognito-idp.us-east-1.amazonaws.com/us-east-1_abcdefghi:app_client_id".

Up to 25 rules can be specified per identity provider.

setIdentityPoolRoles_identityPoolId :: Lens' SetIdentityPoolRoles Text

An identity pool ID in the format REGION:GUID.

setIdentityPoolRoles_roles :: Lens' SetIdentityPoolRoles (HashMap Text Text)

The map of roles associated with this pool. For a given role, the key will be either "authenticated" or "unauthenticated" and the value will be the Role ARN.

Destructuring the Response

data SetIdentityPoolRolesResponse

See: newSetIdentityPoolRolesResponse smart constructor.

Constructors

SetIdentityPoolRolesResponse'

newSetIdentityPoolRolesResponse :: SetIdentityPoolRolesResponse

Create a value of SetIdentityPoolRolesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.