Network.AWS.CognitoIdentity.Types

Description

Service Configuration

defaultService :: Service

API version 2014-06-30 of the Amazon Cognito Identity SDK configuration.

Errors

_InvalidIdentityPoolConfigurationException :: AsError a => Getting (First ServiceError) a ServiceError

Thrown if the identity pool has no role associated for the given auth type (auth/unauth) or if the AssumeRole fails.

_InvalidParameterException :: AsError a => Getting (First ServiceError) a ServiceError

Thrown for missing or bad input parameter(s).

_NotAuthorizedException :: AsError a => Getting (First ServiceError) a ServiceError

Thrown when a user is not authorized to access the requested resource.

_InternalErrorException :: AsError a => Getting (First ServiceError) a ServiceError

Thrown when the service encounters an error during processing the request.

_ExternalServiceException :: AsError a => Getting (First ServiceError) a ServiceError

An exception thrown when a dependent service such as Facebook or Twitter is not responding

_TooManyRequestsException :: AsError a => Getting (First ServiceError) a ServiceError

Thrown when a request is throttled.

_ConcurrentModificationException :: AsError a => Getting (First ServiceError) a ServiceError

Thrown if there are parallel requests to modify a resource.

_ResourceConflictException :: AsError a => Getting (First ServiceError) a ServiceError

Thrown when a user tries to use a login which is already linked to another account.

_DeveloperUserAlreadyRegisteredException :: AsError a => Getting (First ServiceError) a ServiceError

The provided developer user identifier is already registered with Cognito under a different identity ID.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

Thrown when the requested resource (for example, a dataset or record) does not exist.

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Thrown when the total number of user pools has exceeded a preset limit.

AmbiguousRoleResolutionType

newtype AmbiguousRoleResolutionType

Constructors

AmbiguousRoleResolutionType'

Fields

fromAmbiguousRoleResolutionType :: Text

Bundled Patterns

pattern AmbiguousRoleResolutionType_AuthenticatedRole :: AmbiguousRoleResolutionType

pattern AmbiguousRoleResolutionType_Deny :: AmbiguousRoleResolutionType

CognitoErrorCode

newtype CognitoErrorCode

Constructors

CognitoErrorCode'

Fields

fromCognitoErrorCode :: Text

Bundled Patterns

pattern CognitoErrorCode_AccessDenied :: CognitoErrorCode

pattern CognitoErrorCode_InternalServerError :: CognitoErrorCode

MappingRuleMatchType

newtype MappingRuleMatchType

Constructors

MappingRuleMatchType'

Fields

fromMappingRuleMatchType :: Text

Bundled Patterns

pattern MappingRuleMatchType_Contains :: MappingRuleMatchType

pattern MappingRuleMatchType_Equals :: MappingRuleMatchType

pattern MappingRuleMatchType_NotEqual :: MappingRuleMatchType

pattern MappingRuleMatchType_StartsWith :: MappingRuleMatchType

RoleMappingType

newtype RoleMappingType

Constructors

RoleMappingType'

Fields

fromRoleMappingType :: Text

Bundled Patterns

pattern RoleMappingType_Rules :: RoleMappingType

pattern RoleMappingType_Token :: RoleMappingType

CognitoIdentityProvider

data CognitoIdentityProvider

A provider representing an Amazon Cognito user pool and its client ID.

See: newCognitoIdentityProvider smart constructor.

Constructors

CognitoIdentityProvider'

Fields

clientId :: Maybe Text

The client ID for the Amazon Cognito user pool.

serverSideTokenCheck :: Maybe Bool

TRUE if server-side token validation is enabled for the identity provider’s token.

Once you set ServerSideTokenCheck to TRUE for an identity pool, that identity pool will check with the integrated user pools to make sure that the user has not been globally signed out or deleted before the identity pool provides an OIDC token or AWS credentials for the user.

If the user is signed out or deleted, the identity pool will return a 400 Not Authorized error.

providerName :: Maybe Text

The provider name for an Amazon Cognito user pool. For example, cognito-idp.us-east-1.amazonaws.com/us-east-1_123456789.

newCognitoIdentityProvider :: CognitoIdentityProvider

Create a value of CognitoIdentityProvider with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientId:CognitoIdentityProvider', cognitoIdentityProvider_clientId - The client ID for the Amazon Cognito user pool.

$sel:serverSideTokenCheck:CognitoIdentityProvider', cognitoIdentityProvider_serverSideTokenCheck - TRUE if server-side token validation is enabled for the identity provider’s token.

Once you set ServerSideTokenCheck to TRUE for an identity pool, that identity pool will check with the integrated user pools to make sure that the user has not been globally signed out or deleted before the identity pool provides an OIDC token or AWS credentials for the user.

If the user is signed out or deleted, the identity pool will return a 400 Not Authorized error.

$sel:providerName:CognitoIdentityProvider', cognitoIdentityProvider_providerName - The provider name for an Amazon Cognito user pool. For example, cognito-idp.us-east-1.amazonaws.com/us-east-1_123456789.

cognitoIdentityProvider_clientId :: Lens' CognitoIdentityProvider (Maybe Text)

The client ID for the Amazon Cognito user pool.

cognitoIdentityProvider_serverSideTokenCheck :: Lens' CognitoIdentityProvider (Maybe Bool)

TRUE if server-side token validation is enabled for the identity provider’s token.

Once you set ServerSideTokenCheck to TRUE for an identity pool, that identity pool will check with the integrated user pools to make sure that the user has not been globally signed out or deleted before the identity pool provides an OIDC token or AWS credentials for the user.

If the user is signed out or deleted, the identity pool will return a 400 Not Authorized error.

cognitoIdentityProvider_providerName :: Lens' CognitoIdentityProvider (Maybe Text)

The provider name for an Amazon Cognito user pool. For example, cognito-idp.us-east-1.amazonaws.com/us-east-1_123456789.

Credentials

data Credentials

Credentials for the provided identity ID.

See: newCredentials smart constructor.

Constructors

Credentials'

Fields

sessionToken :: Maybe Text

The Session Token portion of the credentials

expiration :: Maybe POSIX

The date at which these credentials will expire.

secretKey :: Maybe Text

The Secret Access Key portion of the credentials

accessKeyId :: Maybe Text

The Access Key portion of the credentials.

newCredentials :: Credentials

Create a value of Credentials with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sessionToken:Credentials', credentials_sessionToken - The Session Token portion of the credentials

$sel:expiration:Credentials', credentials_expiration - The date at which these credentials will expire.

$sel:secretKey:Credentials', credentials_secretKey - The Secret Access Key portion of the credentials

$sel:accessKeyId:Credentials', credentials_accessKeyId - The Access Key portion of the credentials.

credentials_sessionToken :: Lens' Credentials (Maybe Text)

The Session Token portion of the credentials

credentials_expiration :: Lens' Credentials (Maybe UTCTime)

The date at which these credentials will expire.

credentials_secretKey :: Lens' Credentials (Maybe Text)

The Secret Access Key portion of the credentials

credentials_accessKeyId :: Lens' Credentials (Maybe Text)

The Access Key portion of the credentials.

IdentityDescription

data IdentityDescription

A description of the identity.

See: newIdentityDescription smart constructor.

Constructors

IdentityDescription'

Fields

lastModifiedDate :: Maybe POSIX

Date on which the identity was last modified.

creationDate :: Maybe POSIX

Date on which the identity was created.

logins :: Maybe [Text]

The provider names.

identityId :: Maybe Text

A unique identifier in the format REGION:GUID.

newIdentityDescription :: IdentityDescription

Create a value of IdentityDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedDate:IdentityDescription', identityDescription_lastModifiedDate - Date on which the identity was last modified.

$sel:creationDate:IdentityDescription', identityDescription_creationDate - Date on which the identity was created.

$sel:logins:IdentityDescription', identityDescription_logins - The provider names.

$sel:identityId:IdentityDescription', identityDescription_identityId - A unique identifier in the format REGION:GUID.

identityDescription_lastModifiedDate :: Lens' IdentityDescription (Maybe UTCTime)

Date on which the identity was last modified.

identityDescription_creationDate :: Lens' IdentityDescription (Maybe UTCTime)

Date on which the identity was created.

identityDescription_logins :: Lens' IdentityDescription (Maybe [Text])

The provider names.

identityDescription_identityId :: Lens' IdentityDescription (Maybe Text)

A unique identifier in the format REGION:GUID.

IdentityPool

data IdentityPool

An object representing an Amazon Cognito identity pool.

See: newIdentityPool smart constructor.

Constructors

IdentityPool'

Fields

samlProviderARNs :: Maybe [Text]

An array of Amazon Resource Names (ARNs) of the SAML provider for your identity pool.

supportedLoginProviders :: Maybe (HashMap Text Text)

Optional key:value pairs mapping provider names to provider app IDs.

allowClassicFlow :: Maybe Bool

Enables or disables the Basic (Classic) authentication flow. For more information, see Identity Pools (Federated Identities) Authentication Flow in the Amazon Cognito Developer Guide.

developerProviderName :: Maybe Text

The "domain" by which Cognito will refer to your users.

identityPoolTags :: Maybe (HashMap Text Text)

The tags that are assigned to the identity pool. A tag is a label that you can apply to identity pools to categorize and manage them in different ways, such as by purpose, owner, environment, or other criteria.

openIdConnectProviderARNs :: Maybe [Text]

The ARNs of the OpenID Connect providers.

cognitoIdentityProviders :: Maybe [CognitoIdentityProvider]

A list representing an Amazon Cognito user pool and its client ID.

identityPoolId :: Text

An identity pool ID in the format REGION:GUID.

identityPoolName :: Text

A string that you provide.

allowUnauthenticatedIdentities :: Bool

TRUE if the identity pool supports unauthenticated logins.

newIdentityPool

Arguments

:: Text
$sel:identityPoolId:IdentityPool'

-> Text
$sel:identityPoolName:IdentityPool'

-> Bool
$sel:allowUnauthenticatedIdentities:IdentityPool'

-> IdentityPool

Create a value of IdentityPool with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:samlProviderARNs:IdentityPool', identityPool_samlProviderARNs - An array of Amazon Resource Names (ARNs) of the SAML provider for your identity pool.

$sel:supportedLoginProviders:IdentityPool', identityPool_supportedLoginProviders - Optional key:value pairs mapping provider names to provider app IDs.

$sel:allowClassicFlow:IdentityPool', identityPool_allowClassicFlow - Enables or disables the Basic (Classic) authentication flow. For more information, see Identity Pools (Federated Identities) Authentication Flow in the Amazon Cognito Developer Guide.

$sel:developerProviderName:IdentityPool', identityPool_developerProviderName - The "domain" by which Cognito will refer to your users.

$sel:identityPoolTags:IdentityPool', identityPool_identityPoolTags - The tags that are assigned to the identity pool. A tag is a label that you can apply to identity pools to categorize and manage them in different ways, such as by purpose, owner, environment, or other criteria.

$sel:openIdConnectProviderARNs:IdentityPool', identityPool_openIdConnectProviderARNs - The ARNs of the OpenID Connect providers.

$sel:cognitoIdentityProviders:IdentityPool', identityPool_cognitoIdentityProviders - A list representing an Amazon Cognito user pool and its client ID.

$sel:identityPoolId:IdentityPool', identityPool_identityPoolId - An identity pool ID in the format REGION:GUID.

$sel:identityPoolName:IdentityPool', identityPool_identityPoolName - A string that you provide.

$sel:allowUnauthenticatedIdentities:IdentityPool', identityPool_allowUnauthenticatedIdentities - TRUE if the identity pool supports unauthenticated logins.

identityPool_samlProviderARNs :: Lens' IdentityPool (Maybe [Text])

An array of Amazon Resource Names (ARNs) of the SAML provider for your identity pool.

identityPool_supportedLoginProviders :: Lens' IdentityPool (Maybe (HashMap Text Text))

Optional key:value pairs mapping provider names to provider app IDs.

identityPool_allowClassicFlow :: Lens' IdentityPool (Maybe Bool)

Enables or disables the Basic (Classic) authentication flow. For more information, see Identity Pools (Federated Identities) Authentication Flow in the Amazon Cognito Developer Guide.

identityPool_developerProviderName :: Lens' IdentityPool (Maybe Text)

The "domain" by which Cognito will refer to your users.

identityPool_identityPoolTags :: Lens' IdentityPool (Maybe (HashMap Text Text))

The tags that are assigned to the identity pool. A tag is a label that you can apply to identity pools to categorize and manage them in different ways, such as by purpose, owner, environment, or other criteria.

identityPool_openIdConnectProviderARNs :: Lens' IdentityPool (Maybe [Text])

The ARNs of the OpenID Connect providers.

identityPool_cognitoIdentityProviders :: Lens' IdentityPool (Maybe [CognitoIdentityProvider])

A list representing an Amazon Cognito user pool and its client ID.

identityPool_identityPoolId :: Lens' IdentityPool Text

An identity pool ID in the format REGION:GUID.

identityPool_identityPoolName :: Lens' IdentityPool Text

A string that you provide.

identityPool_allowUnauthenticatedIdentities :: Lens' IdentityPool Bool

TRUE if the identity pool supports unauthenticated logins.

IdentityPoolShortDescription

data IdentityPoolShortDescription

A description of the identity pool.

See: newIdentityPoolShortDescription smart constructor.

Constructors

IdentityPoolShortDescription'

Fields

identityPoolId :: Maybe Text

An identity pool ID in the format REGION:GUID.

identityPoolName :: Maybe Text

A string that you provide.

newIdentityPoolShortDescription :: IdentityPoolShortDescription

Create a value of IdentityPoolShortDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identityPoolId:IdentityPoolShortDescription', identityPoolShortDescription_identityPoolId - An identity pool ID in the format REGION:GUID.

$sel:identityPoolName:IdentityPoolShortDescription', identityPoolShortDescription_identityPoolName - A string that you provide.

identityPoolShortDescription_identityPoolId :: Lens' IdentityPoolShortDescription (Maybe Text)

An identity pool ID in the format REGION:GUID.

identityPoolShortDescription_identityPoolName :: Lens' IdentityPoolShortDescription (Maybe Text)

A string that you provide.

MappingRule

data MappingRule

A rule that maps a claim name, a claim value, and a match type to a role ARN.

See: newMappingRule smart constructor.

Constructors

MappingRule'

Fields

claim :: Text

The claim name that must be present in the token, for example, "isAdmin" or "paid".

matchType :: MappingRuleMatchType

The match condition that specifies how closely the claim value in the IdP token must match Value.

value :: Text

A brief string that the claim must match, for example, "paid" or "yes".

roleARN :: Text

The role ARN.

newMappingRule

Arguments

:: Text
$sel:claim:MappingRule'

-> MappingRuleMatchType
$sel:matchType:MappingRule'

-> Text
$sel:value:MappingRule'

-> Text
$sel:roleARN:MappingRule'

-> MappingRule

Create a value of MappingRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:claim:MappingRule', mappingRule_claim - The claim name that must be present in the token, for example, "isAdmin" or "paid".

$sel:matchType:MappingRule', mappingRule_matchType - The match condition that specifies how closely the claim value in the IdP token must match Value.

$sel:value:MappingRule', mappingRule_value - A brief string that the claim must match, for example, "paid" or "yes".

$sel:roleARN:MappingRule', mappingRule_roleARN - The role ARN.

mappingRule_claim :: Lens' MappingRule Text

The claim name that must be present in the token, for example, "isAdmin" or "paid".

mappingRule_matchType :: Lens' MappingRule MappingRuleMatchType

The match condition that specifies how closely the claim value in the IdP token must match Value.

mappingRule_value :: Lens' MappingRule Text

A brief string that the claim must match, for example, "paid" or "yes".

mappingRule_roleARN :: Lens' MappingRule Text

The role ARN.

RoleMapping

data RoleMapping

A role mapping.

See: newRoleMapping smart constructor.

Constructors

RoleMapping'

Fields

rulesConfiguration :: Maybe RulesConfigurationType

The rules to be used for mapping users to roles.

If you specify Rules as the role mapping type, RulesConfiguration is required.

ambiguousRoleResolution :: Maybe AmbiguousRoleResolutionType

If you specify Token or Rules as the Type, AmbiguousRoleResolution is required.

Specifies the action to be taken if either no rules match the claim value for the Rules type, or there is no cognito:preferred_role claim and there are multiple cognito:roles matches for the Token type.

type' :: RoleMappingType

The role mapping type. Token will use cognito:roles and cognito:preferred_role claims from the Cognito identity provider token to map groups to roles. Rules will attempt to match claims from the token to map to a role.

newRoleMapping

Arguments

:: RoleMappingType
$sel:type':RoleMapping'

-> RoleMapping

Create a value of RoleMapping with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:rulesConfiguration:RoleMapping', roleMapping_rulesConfiguration - The rules to be used for mapping users to roles.

If you specify Rules as the role mapping type, RulesConfiguration is required.

$sel:ambiguousRoleResolution:RoleMapping', roleMapping_ambiguousRoleResolution - If you specify Token or Rules as the Type, AmbiguousRoleResolution is required.

Specifies the action to be taken if either no rules match the claim value for the Rules type, or there is no cognito:preferred_role claim and there are multiple cognito:roles matches for the Token type.

$sel:type':RoleMapping', roleMapping_type - The role mapping type. Token will use cognito:roles and cognito:preferred_role claims from the Cognito identity provider token to map groups to roles. Rules will attempt to match claims from the token to map to a role.

roleMapping_rulesConfiguration :: Lens' RoleMapping (Maybe RulesConfigurationType)

The rules to be used for mapping users to roles.

If you specify Rules as the role mapping type, RulesConfiguration is required.

roleMapping_ambiguousRoleResolution :: Lens' RoleMapping (Maybe AmbiguousRoleResolutionType)

If you specify Token or Rules as the Type, AmbiguousRoleResolution is required.

Specifies the action to be taken if either no rules match the claim value for the Rules type, or there is no cognito:preferred_role claim and there are multiple cognito:roles matches for the Token type.

roleMapping_type :: Lens' RoleMapping RoleMappingType

The role mapping type. Token will use cognito:roles and cognito:preferred_role claims from the Cognito identity provider token to map groups to roles. Rules will attempt to match claims from the token to map to a role.

RulesConfigurationType

data RulesConfigurationType

A container for rules.

See: newRulesConfigurationType smart constructor.

Constructors

RulesConfigurationType'

Fields

rules :: NonEmpty MappingRule

An array of rules. You can specify up to 25 rules per identity provider.

Rules are evaluated in order. The first one to match specifies the role.

newRulesConfigurationType

Arguments

:: NonEmpty MappingRule
$sel:rules:RulesConfigurationType'

-> RulesConfigurationType

Create a value of RulesConfigurationType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:rules:RulesConfigurationType', rulesConfigurationType_rules - An array of rules. You can specify up to 25 rules per identity provider.

Rules are evaluated in order. The first one to match specifies the role.

rulesConfigurationType_rules :: Lens' RulesConfigurationType (NonEmpty MappingRule)

An array of rules. You can specify up to 25 rules per identity provider.

Rules are evaluated in order. The first one to match specifies the role.

UnprocessedIdentityId

data UnprocessedIdentityId

An array of UnprocessedIdentityId objects, each of which contains an ErrorCode and IdentityId.

See: newUnprocessedIdentityId smart constructor.

Constructors

UnprocessedIdentityId'

Fields

errorCode :: Maybe CognitoErrorCode

The error code indicating the type of error that occurred.

identityId :: Maybe Text

A unique identifier in the format REGION:GUID.

newUnprocessedIdentityId :: UnprocessedIdentityId

Create a value of UnprocessedIdentityId with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:errorCode:UnprocessedIdentityId', unprocessedIdentityId_errorCode - The error code indicating the type of error that occurred.

$sel:identityId:UnprocessedIdentityId', unprocessedIdentityId_identityId - A unique identifier in the format REGION:GUID.

unprocessedIdentityId_errorCode :: Lens' UnprocessedIdentityId (Maybe CognitoErrorCode)

The error code indicating the type of error that occurred.

unprocessedIdentityId_identityId :: Lens' UnprocessedIdentityId (Maybe Text)

A unique identifier in the format REGION:GUID.