Network.AWS.CognitoIdentity.UnlinkDeveloperIdentity

Description

Unlinks a DeveloperUserIdentifier from an existing identity. Unlinked developer users will be considered new identities next time they are seen. If, for a given Cognito identity, you remove all federated identities as well as the developer user identifier, the Cognito identity becomes inaccessible.

You must use AWS Developer credentials to call this API.

Creating a Request

data UnlinkDeveloperIdentity

Input to the UnlinkDeveloperIdentity action.

See: newUnlinkDeveloperIdentity smart constructor.

Constructors

UnlinkDeveloperIdentity'

Fields

identityId :: Text

A unique identifier in the format REGION:GUID.

identityPoolId :: Text

An identity pool ID in the format REGION:GUID.

developerProviderName :: Text

The "domain" by which Cognito will refer to your users.

developerUserIdentifier :: Text

A unique ID used by your backend authentication process to identify a user.

newUnlinkDeveloperIdentity

Arguments

:: Text
$sel:identityId:UnlinkDeveloperIdentity'

-> Text
$sel:identityPoolId:UnlinkDeveloperIdentity'

-> Text
$sel:developerProviderName:UnlinkDeveloperIdentity'

-> Text
$sel:developerUserIdentifier:UnlinkDeveloperIdentity'

-> UnlinkDeveloperIdentity

Create a value of UnlinkDeveloperIdentity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identityId:UnlinkDeveloperIdentity', unlinkDeveloperIdentity_identityId - A unique identifier in the format REGION:GUID.

$sel:identityPoolId:UnlinkDeveloperIdentity', unlinkDeveloperIdentity_identityPoolId - An identity pool ID in the format REGION:GUID.

$sel:developerProviderName:UnlinkDeveloperIdentity', unlinkDeveloperIdentity_developerProviderName - The "domain" by which Cognito will refer to your users.

$sel:developerUserIdentifier:UnlinkDeveloperIdentity', unlinkDeveloperIdentity_developerUserIdentifier - A unique ID used by your backend authentication process to identify a user.

Request Lenses

unlinkDeveloperIdentity_identityId :: Lens' UnlinkDeveloperIdentity Text

A unique identifier in the format REGION:GUID.

unlinkDeveloperIdentity_identityPoolId :: Lens' UnlinkDeveloperIdentity Text

An identity pool ID in the format REGION:GUID.

unlinkDeveloperIdentity_developerProviderName :: Lens' UnlinkDeveloperIdentity Text

The "domain" by which Cognito will refer to your users.

unlinkDeveloperIdentity_developerUserIdentifier :: Lens' UnlinkDeveloperIdentity Text

A unique ID used by your backend authentication process to identify a user.

Destructuring the Response

data UnlinkDeveloperIdentityResponse

See: newUnlinkDeveloperIdentityResponse smart constructor.

Constructors

UnlinkDeveloperIdentityResponse'

newUnlinkDeveloperIdentityResponse :: UnlinkDeveloperIdentityResponse

Create a value of UnlinkDeveloperIdentityResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.