Network.AWS.CognitoIdentity.UpdateIdentityPool

Description

Updates an identity pool.

You must use AWS Developer credentials to call this API.

Creating a Request

data UpdateIdentityPool

An object representing an Amazon Cognito identity pool.

See: newUpdateIdentityPool smart constructor.

Constructors

UpdateIdentityPool'

Fields

samlProviderARNs :: Maybe [Text]

An array of Amazon Resource Names (ARNs) of the SAML provider for your identity pool.

supportedLoginProviders :: Maybe (HashMap Text Text)

Optional key:value pairs mapping provider names to provider app IDs.

allowClassicFlow :: Maybe Bool

Enables or disables the Basic (Classic) authentication flow. For more information, see Identity Pools (Federated Identities) Authentication Flow in the Amazon Cognito Developer Guide.

developerProviderName :: Maybe Text

The "domain" by which Cognito will refer to your users.

identityPoolTags :: Maybe (HashMap Text Text)

The tags that are assigned to the identity pool. A tag is a label that you can apply to identity pools to categorize and manage them in different ways, such as by purpose, owner, environment, or other criteria.

openIdConnectProviderARNs :: Maybe [Text]

The ARNs of the OpenID Connect providers.

cognitoIdentityProviders :: Maybe [CognitoIdentityProvider]

A list representing an Amazon Cognito user pool and its client ID.

identityPoolId :: Text

An identity pool ID in the format REGION:GUID.

identityPoolName :: Text

A string that you provide.

allowUnauthenticatedIdentities :: Bool

TRUE if the identity pool supports unauthenticated logins.

newUpdateIdentityPool

Arguments

:: Text
$sel:identityPoolId:UpdateIdentityPool'

-> Text
$sel:identityPoolName:UpdateIdentityPool'

-> Bool
$sel:allowUnauthenticatedIdentities:UpdateIdentityPool'

-> UpdateIdentityPool

Create a value of UpdateIdentityPool with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:samlProviderARNs:UpdateIdentityPool', updateIdentityPool_samlProviderARNs - An array of Amazon Resource Names (ARNs) of the SAML provider for your identity pool.

$sel:supportedLoginProviders:UpdateIdentityPool', updateIdentityPool_supportedLoginProviders - Optional key:value pairs mapping provider names to provider app IDs.

$sel:allowClassicFlow:UpdateIdentityPool', updateIdentityPool_allowClassicFlow - Enables or disables the Basic (Classic) authentication flow. For more information, see Identity Pools (Federated Identities) Authentication Flow in the Amazon Cognito Developer Guide.

$sel:developerProviderName:UpdateIdentityPool', updateIdentityPool_developerProviderName - The "domain" by which Cognito will refer to your users.

$sel:identityPoolTags:UpdateIdentityPool', updateIdentityPool_identityPoolTags - The tags that are assigned to the identity pool. A tag is a label that you can apply to identity pools to categorize and manage them in different ways, such as by purpose, owner, environment, or other criteria.

$sel:openIdConnectProviderARNs:UpdateIdentityPool', updateIdentityPool_openIdConnectProviderARNs - The ARNs of the OpenID Connect providers.

$sel:cognitoIdentityProviders:UpdateIdentityPool', updateIdentityPool_cognitoIdentityProviders - A list representing an Amazon Cognito user pool and its client ID.

$sel:identityPoolId:UpdateIdentityPool', updateIdentityPool_identityPoolId - An identity pool ID in the format REGION:GUID.

$sel:identityPoolName:UpdateIdentityPool', updateIdentityPool_identityPoolName - A string that you provide.

$sel:allowUnauthenticatedIdentities:UpdateIdentityPool', updateIdentityPool_allowUnauthenticatedIdentities - TRUE if the identity pool supports unauthenticated logins.

Request Lenses

updateIdentityPool_samlProviderARNs :: Lens' UpdateIdentityPool (Maybe [Text])

An array of Amazon Resource Names (ARNs) of the SAML provider for your identity pool.

updateIdentityPool_supportedLoginProviders :: Lens' UpdateIdentityPool (Maybe (HashMap Text Text))

Optional key:value pairs mapping provider names to provider app IDs.

updateIdentityPool_allowClassicFlow :: Lens' UpdateIdentityPool (Maybe Bool)

Enables or disables the Basic (Classic) authentication flow. For more information, see Identity Pools (Federated Identities) Authentication Flow in the Amazon Cognito Developer Guide.

updateIdentityPool_developerProviderName :: Lens' UpdateIdentityPool (Maybe Text)

The "domain" by which Cognito will refer to your users.

updateIdentityPool_identityPoolTags :: Lens' UpdateIdentityPool (Maybe (HashMap Text Text))

The tags that are assigned to the identity pool. A tag is a label that you can apply to identity pools to categorize and manage them in different ways, such as by purpose, owner, environment, or other criteria.

updateIdentityPool_openIdConnectProviderARNs :: Lens' UpdateIdentityPool (Maybe [Text])

The ARNs of the OpenID Connect providers.

updateIdentityPool_cognitoIdentityProviders :: Lens' UpdateIdentityPool (Maybe [CognitoIdentityProvider])

A list representing an Amazon Cognito user pool and its client ID.

updateIdentityPool_identityPoolId :: Lens' UpdateIdentityPool Text

An identity pool ID in the format REGION:GUID.

updateIdentityPool_identityPoolName :: Lens' UpdateIdentityPool Text

A string that you provide.

updateIdentityPool_allowUnauthenticatedIdentities :: Lens' UpdateIdentityPool Bool

TRUE if the identity pool supports unauthenticated logins.

Destructuring the Response

data IdentityPool

An object representing an Amazon Cognito identity pool.

See: newIdentityPool smart constructor.

Constructors

IdentityPool'

Fields

samlProviderARNs :: Maybe [Text]

An array of Amazon Resource Names (ARNs) of the SAML provider for your identity pool.

supportedLoginProviders :: Maybe (HashMap Text Text)

Optional key:value pairs mapping provider names to provider app IDs.

allowClassicFlow :: Maybe Bool

Enables or disables the Basic (Classic) authentication flow. For more information, see Identity Pools (Federated Identities) Authentication Flow in the Amazon Cognito Developer Guide.

developerProviderName :: Maybe Text

The "domain" by which Cognito will refer to your users.

identityPoolTags :: Maybe (HashMap Text Text)

The tags that are assigned to the identity pool. A tag is a label that you can apply to identity pools to categorize and manage them in different ways, such as by purpose, owner, environment, or other criteria.

openIdConnectProviderARNs :: Maybe [Text]

The ARNs of the OpenID Connect providers.

cognitoIdentityProviders :: Maybe [CognitoIdentityProvider]

A list representing an Amazon Cognito user pool and its client ID.

identityPoolId :: Text

An identity pool ID in the format REGION:GUID.

identityPoolName :: Text

A string that you provide.

allowUnauthenticatedIdentities :: Bool

TRUE if the identity pool supports unauthenticated logins.

newIdentityPool

Arguments

:: Text
$sel:identityPoolId:IdentityPool'

-> Text
$sel:identityPoolName:IdentityPool'

-> Bool
$sel:allowUnauthenticatedIdentities:IdentityPool'

-> IdentityPool

Create a value of IdentityPool with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:samlProviderARNs:IdentityPool', identityPool_samlProviderARNs - An array of Amazon Resource Names (ARNs) of the SAML provider for your identity pool.

$sel:supportedLoginProviders:IdentityPool', identityPool_supportedLoginProviders - Optional key:value pairs mapping provider names to provider app IDs.

$sel:allowClassicFlow:IdentityPool', identityPool_allowClassicFlow - Enables or disables the Basic (Classic) authentication flow. For more information, see Identity Pools (Federated Identities) Authentication Flow in the Amazon Cognito Developer Guide.

$sel:developerProviderName:IdentityPool', identityPool_developerProviderName - The "domain" by which Cognito will refer to your users.

$sel:identityPoolTags:IdentityPool', identityPool_identityPoolTags - The tags that are assigned to the identity pool. A tag is a label that you can apply to identity pools to categorize and manage them in different ways, such as by purpose, owner, environment, or other criteria.

$sel:openIdConnectProviderARNs:IdentityPool', identityPool_openIdConnectProviderARNs - The ARNs of the OpenID Connect providers.

$sel:cognitoIdentityProviders:IdentityPool', identityPool_cognitoIdentityProviders - A list representing an Amazon Cognito user pool and its client ID.

$sel:identityPoolId:IdentityPool', identityPool_identityPoolId - An identity pool ID in the format REGION:GUID.

$sel:identityPoolName:IdentityPool', identityPool_identityPoolName - A string that you provide.

$sel:allowUnauthenticatedIdentities:IdentityPool', identityPool_allowUnauthenticatedIdentities - TRUE if the identity pool supports unauthenticated logins.

Response Lenses

identityPool_samlProviderARNs :: Lens' IdentityPool (Maybe [Text])

An array of Amazon Resource Names (ARNs) of the SAML provider for your identity pool.

identityPool_supportedLoginProviders :: Lens' IdentityPool (Maybe (HashMap Text Text))

Optional key:value pairs mapping provider names to provider app IDs.

identityPool_allowClassicFlow :: Lens' IdentityPool (Maybe Bool)

Enables or disables the Basic (Classic) authentication flow. For more information, see Identity Pools (Federated Identities) Authentication Flow in the Amazon Cognito Developer Guide.

identityPool_developerProviderName :: Lens' IdentityPool (Maybe Text)

The "domain" by which Cognito will refer to your users.

identityPool_identityPoolTags :: Lens' IdentityPool (Maybe (HashMap Text Text))

The tags that are assigned to the identity pool. A tag is a label that you can apply to identity pools to categorize and manage them in different ways, such as by purpose, owner, environment, or other criteria.

identityPool_openIdConnectProviderARNs :: Lens' IdentityPool (Maybe [Text])

The ARNs of the OpenID Connect providers.

identityPool_cognitoIdentityProviders :: Lens' IdentityPool (Maybe [CognitoIdentityProvider])

A list representing an Amazon Cognito user pool and its client ID.

identityPool_identityPoolId :: Lens' IdentityPool Text

An identity pool ID in the format REGION:GUID.

identityPool_identityPoolName :: Lens' IdentityPool Text

A string that you provide.

identityPool_allowUnauthenticatedIdentities :: Lens' IdentityPool Bool

TRUE if the identity pool supports unauthenticated logins.