Network.AWS.CognitoIdentity.V2014_06_30

Description

Amazon Cognito is a web service that facilitates the delivery of scoped, temporary credentials to mobile devices or other untrusted environments. Amazon Cognito uniquely identifies a device or user and supplies the user with a consistent identity throughout the lifetime of an application. Amazon Cognito lets users authenticate with third-party identity providers (Facebook, Google, or Login with Amazon). As a developer, you decide which identity providers to trust. You can also choose to support unauthenticated access from your application. Your users are provided with Cognito tokens that uniquely identify their device and any information provided about third-party logins.

Documentation

createIdentityPool

Arguments

:: Text
_cipiIdentityPoolName

-> Bool
_cipiAllowUnauthenticatedIdentities

-> CreateIdentityPool

Minimum specification for a CreateIdentityPool request.

data CreateIdentityPool

Constructors

CreateIdentityPool

Fields

_cipiIdentityPoolName :: Text

_cipiAllowUnauthenticatedIdentities :: Bool

_cipiSupportedLoginProviders :: HashMap Text Text

cipiSupportedLoginProviders :: Lens' CreateIdentityPool (HashMap Text Text)

cipiIdentityPoolName :: Lens' CreateIdentityPool Text

cipiAllowUnauthenticatedIdentities :: Lens' CreateIdentityPool Bool

data CreateIdentityPoolResponse

Constructors

CreateIdentityPoolResponse

Fields

_iwIdentityPoolId :: Text

_iwIdentityPoolName :: Text

_iwAllowUnauthenticatedIdentities :: Bool

_iwSupportedLoginProviders :: HashMap Text Text

iwSupportedLoginProviders :: Lens' CreateIdentityPoolResponse (HashMap Text Text)

iwIdentityPoolName :: Lens' CreateIdentityPoolResponse Text

iwIdentityPoolId :: Lens' CreateIdentityPoolResponse Text

iwAllowUnauthenticatedIdentities :: Lens' CreateIdentityPoolResponse Bool

data DeleteIdentityPool

Constructors

DeleteIdentityPool

Fields

_dipjIdentityPoolId :: Text

dipjIdentityPoolId :: Iso' DeleteIdentityPool Text

data DeleteIdentityPoolResponse

Constructors

DeleteIdentityPoolResponse

data DescribeIdentityPool

Constructors

DescribeIdentityPool

Fields

_dipiIdentityPoolId :: Text

dipiIdentityPoolId :: Iso' DescribeIdentityPool Text

data DescribeIdentityPoolResponse

Constructors

DescribeIdentityPoolResponse

Fields

_ipIdentityPoolId :: Text

_ipIdentityPoolName :: Text

_ipAllowUnauthenticatedIdentities :: Bool

_ipSupportedLoginProviders :: HashMap Text Text

ipSupportedLoginProviders :: Lens' DescribeIdentityPoolResponse (HashMap Text Text)

ipIdentityPoolName :: Lens' DescribeIdentityPoolResponse Text

ipIdentityPoolId :: Lens' DescribeIdentityPoolResponse Text

ipAllowUnauthenticatedIdentities :: Lens' DescribeIdentityPoolResponse Bool

getId

Arguments

:: Text
_giiAccountId

-> Text
_giiIdentityPoolId

-> GetId

Minimum specification for a GetId request.

data GetId

Constructors

GetId

Fields

_giiAccountId :: Text

_giiIdentityPoolId :: Text

_giiLogins :: HashMap Text Text

giiLogins :: Lens' GetId (HashMap Text Text)

giiIdentityPoolId :: Lens' GetId Text

giiAccountId :: Lens' GetId Text

data GetIdResponse

Constructors

GetIdResponse

Fields

_girIdentityId :: Maybe Text

girIdentityId :: Iso' GetIdResponse (Maybe Text)

getOpenIdToken

Arguments

:: Text
_goitiIdentityId

-> GetOpenIdToken

Minimum specification for a GetOpenIdToken request.

data GetOpenIdToken

Constructors

GetOpenIdToken

Fields

_goitiIdentityId :: Text

_goitiLogins :: HashMap Text Text

goitiLogins :: Lens' GetOpenIdToken (HashMap Text Text)

goitiIdentityId :: Lens' GetOpenIdToken Text

data GetOpenIdTokenResponse

Constructors

GetOpenIdTokenResponse

Fields

_goitrIdentityId :: Maybe Text

_goitrToken :: Maybe Text

goitrToken :: Lens' GetOpenIdTokenResponse (Maybe Text)

goitrIdentityId :: Lens' GetOpenIdTokenResponse (Maybe Text)

listIdentities

Arguments

:: Text
_liiIdentityPoolId

-> ListIdentities

Minimum specification for a ListIdentities request.

data ListIdentities

Constructors

ListIdentities

Fields

_liiIdentityPoolId :: Text

_liiNextToken :: Maybe Text

_liiMaxResults :: Maybe Integer

liiNextToken :: Lens' ListIdentities (Maybe Text)

liiMaxResults :: Lens' ListIdentities (Maybe Integer)

liiIdentityPoolId :: Lens' ListIdentities Text

data ListIdentitiesResponse

Constructors

ListIdentitiesResponse

Fields

_lirIdentities :: [IdentityDescription]

_lirIdentityPoolId :: Maybe Text

_lirNextToken :: Maybe Text

lirNextToken :: Lens' ListIdentitiesResponse (Maybe Text)

lirIdentityPoolId :: Lens' ListIdentitiesResponse (Maybe Text)

lirIdentities :: Lens' ListIdentitiesResponse [IdentityDescription]

listIdentityPools :: ListIdentityPools

Minimum specification for a ListIdentityPools request.

data ListIdentityPools

Constructors

ListIdentityPools

Fields

_lipiNextToken :: Maybe Text

_lipiMaxResults :: Maybe Integer

lipiNextToken :: Lens' ListIdentityPools (Maybe Text)

lipiMaxResults :: Lens' ListIdentityPools (Maybe Integer)

data ListIdentityPoolsResponse

Constructors

ListIdentityPoolsResponse

Fields

_liprIdentityPools :: [IdentityPoolShortDescription]

_liprNextToken :: Maybe Text

liprNextToken :: Lens' ListIdentityPoolsResponse (Maybe Text)

liprIdentityPools :: Lens' ListIdentityPoolsResponse [IdentityPoolShortDescription]

data CognitoIdentity

Supported version (2014-06-30) of the Amazon Cognito Identity Service service.

data IdentityDescription

Constructors

IdentityDescription

Fields

_idLogins :: [Text]

_idIdentityId :: Maybe Text

data IdentityPoolShortDescription

Constructors

IdentityPoolShortDescription

Fields

_ipsdIdentityPoolId :: Maybe Text

_ipsdIdentityPoolName :: Maybe Text

idLogins :: Lens' IdentityDescription [Text]

idIdentityId :: Lens' IdentityDescription (Maybe Text)

ipsdIdentityPoolName :: Lens' IdentityPoolShortDescription (Maybe Text)

ipsdIdentityPoolId :: Lens' IdentityPoolShortDescription (Maybe Text)

data UnlinkIdentity

Constructors

UnlinkIdentity

Fields

_uiiIdentityId :: Text

_uiiLoginsToRemove :: [Text]

_uiiLogins :: HashMap Text Text

uiiLoginsToRemove :: Lens' UnlinkIdentity [Text]

uiiLogins :: Lens' UnlinkIdentity (HashMap Text Text)

uiiIdentityId :: Lens' UnlinkIdentity Text

data UnlinkIdentityResponse

Constructors

UnlinkIdentityResponse

updateIdentityPool

Arguments

:: Text
_irIdentityPoolId

-> Text
_irIdentityPoolName

-> Bool
_irAllowUnauthenticatedIdentities

-> UpdateIdentityPool

Minimum specification for a UpdateIdentityPool request.

data UpdateIdentityPool

Constructors

UpdateIdentityPool

Fields

_irIdentityPoolId :: Text

_irIdentityPoolName :: Text

_irAllowUnauthenticatedIdentities :: Bool

_irSupportedLoginProviders :: HashMap Text Text

irSupportedLoginProviders :: Lens' UpdateIdentityPool (HashMap Text Text)

irIdentityPoolName :: Lens' UpdateIdentityPool Text

irIdentityPoolId :: Lens' UpdateIdentityPool Text

irAllowUnauthenticatedIdentities :: Lens' UpdateIdentityPool Bool

data UpdateIdentityPoolResponse

Constructors

UpdateIdentityPoolResponse

Fields

_itIdentityPoolId :: Text

_itIdentityPoolName :: Text

_itAllowUnauthenticatedIdentities :: Bool

_itSupportedLoginProviders :: HashMap Text Text

itSupportedLoginProviders :: Lens' UpdateIdentityPoolResponse (HashMap Text Text)

itIdentityPoolName :: Lens' UpdateIdentityPoolResponse Text

itIdentityPoolId :: Lens' UpdateIdentityPoolResponse Text

itAllowUnauthenticatedIdentities :: Lens' UpdateIdentityPoolResponse Bool