Network.AWS.CognitoIdentity.V2014_06_30.Types

Description

Amazon Cognito is a web service that facilitates the delivery of scoped, temporary credentials to mobile devices or other untrusted environments. Amazon Cognito uniquely identifies a device or user and supplies the user with a consistent identity throughout the lifetime of an application. Amazon Cognito lets users authenticate with third-party identity providers (Facebook, Google, or Login with Amazon). As a developer, you decide which identity providers to trust. You can also choose to support unauthenticated access from your application. Your users are provided with Cognito tokens that uniquely identify their device and any information provided about third-party logins.

Documentation

data CognitoIdentity

Supported version (2014-06-30) of the Amazon Cognito Identity Service service.

data IdentityDescription

Constructors

IdentityDescription

Fields

_idLogins :: [Text]

_idIdentityId :: Maybe Text

data IdentityPoolShortDescription

Constructors

IdentityPoolShortDescription

Fields

_ipsdIdentityPoolId :: Maybe Text

_ipsdIdentityPoolName :: Maybe Text

idLogins :: Lens' IdentityDescription [Text]

idIdentityId :: Lens' IdentityDescription (Maybe Text)

ipsdIdentityPoolName :: Lens' IdentityPoolShortDescription (Maybe Text)

ipsdIdentityPoolId :: Lens' IdentityPoolShortDescription (Maybe Text)