Network.AWS.CognitoIdentityProvider.AdminConfirmSignUp

Description

Confirms user registration as an admin without using a confirmation code. Works on any user.

Calling this action requires developer credentials.

Creating a Request

data AdminConfirmSignUp

Represents the request to confirm user registration.

See: newAdminConfirmSignUp smart constructor.

Constructors

AdminConfirmSignUp'

Fields

clientMetadata :: Maybe (HashMap Text Text)

A map of custom key-value pairs that you can provide as input for any custom workflows that this action triggers.

If your user pool configuration includes triggers, the AdminConfirmSignUp API action invokes the Lambda function that is specified for the post confirmation trigger. When Amazon Cognito invokes this function, it passes a JSON payload, which the function receives as input. In this payload, the clientMetadata attribute provides the data that you assigned to the ClientMetadata parameter in your AdminConfirmSignUp request. In your function code in Lambda, you can process the ClientMetadata value to enhance your workflow for your specific needs.

For more information, see Customizing User Pool Workflows with Lambda Triggers in the Amazon Cognito Developer Guide.

Take the following limitations into consideration when you use the ClientMetadata parameter:

Amazon Cognito does not store the ClientMetadata value. This data is available only to Lambda triggers that are assigned to a user pool to support custom workflows. If your user pool configuration does not include triggers, the ClientMetadata parameter serves no purpose.

Amazon Cognito does not validate the ClientMetadata value.

Amazon Cognito does not encrypt the the ClientMetadata value, so don't use it to provide sensitive information.

userPoolId :: Text

The user pool ID for which you want to confirm user registration.

username :: Sensitive Text

The user name for which you want to confirm user registration.

newAdminConfirmSignUp

Arguments

:: Text
$sel:userPoolId:AdminConfirmSignUp'

-> Text
$sel:username:AdminConfirmSignUp'

-> AdminConfirmSignUp

Create a value of AdminConfirmSignUp with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientMetadata:AdminConfirmSignUp', adminConfirmSignUp_clientMetadata - A map of custom key-value pairs that you can provide as input for any custom workflows that this action triggers.

If your user pool configuration includes triggers, the AdminConfirmSignUp API action invokes the Lambda function that is specified for the post confirmation trigger. When Amazon Cognito invokes this function, it passes a JSON payload, which the function receives as input. In this payload, the clientMetadata attribute provides the data that you assigned to the ClientMetadata parameter in your AdminConfirmSignUp request. In your function code in Lambda, you can process the ClientMetadata value to enhance your workflow for your specific needs.

For more information, see Customizing User Pool Workflows with Lambda Triggers in the Amazon Cognito Developer Guide.

Take the following limitations into consideration when you use the ClientMetadata parameter:

Amazon Cognito does not store the ClientMetadata value. This data is available only to Lambda triggers that are assigned to a user pool to support custom workflows. If your user pool configuration does not include triggers, the ClientMetadata parameter serves no purpose.

Amazon Cognito does not validate the ClientMetadata value.

Amazon Cognito does not encrypt the the ClientMetadata value, so don't use it to provide sensitive information.

$sel:userPoolId:AdminConfirmSignUp', adminConfirmSignUp_userPoolId - The user pool ID for which you want to confirm user registration.

$sel:username:AdminConfirmSignUp', adminConfirmSignUp_username - The user name for which you want to confirm user registration.

Request Lenses

adminConfirmSignUp_clientMetadata :: Lens' AdminConfirmSignUp (Maybe (HashMap Text Text))

A map of custom key-value pairs that you can provide as input for any custom workflows that this action triggers.

If your user pool configuration includes triggers, the AdminConfirmSignUp API action invokes the Lambda function that is specified for the post confirmation trigger. When Amazon Cognito invokes this function, it passes a JSON payload, which the function receives as input. In this payload, the clientMetadata attribute provides the data that you assigned to the ClientMetadata parameter in your AdminConfirmSignUp request. In your function code in Lambda, you can process the ClientMetadata value to enhance your workflow for your specific needs.

For more information, see Customizing User Pool Workflows with Lambda Triggers in the Amazon Cognito Developer Guide.

Take the following limitations into consideration when you use the ClientMetadata parameter:

Amazon Cognito does not store the ClientMetadata value. This data is available only to Lambda triggers that are assigned to a user pool to support custom workflows. If your user pool configuration does not include triggers, the ClientMetadata parameter serves no purpose.

Amazon Cognito does not validate the ClientMetadata value.

Amazon Cognito does not encrypt the the ClientMetadata value, so don't use it to provide sensitive information.

adminConfirmSignUp_userPoolId :: Lens' AdminConfirmSignUp Text

The user pool ID for which you want to confirm user registration.

adminConfirmSignUp_username :: Lens' AdminConfirmSignUp Text

The user name for which you want to confirm user registration.

Destructuring the Response

data AdminConfirmSignUpResponse

Represents the response from the server for the request to confirm registration.

See: newAdminConfirmSignUpResponse smart constructor.

Constructors

AdminConfirmSignUpResponse'

Fields

httpStatus :: Int

The response's http status code.

newAdminConfirmSignUpResponse

Arguments

:: Int
$sel:httpStatus:AdminConfirmSignUpResponse'

-> AdminConfirmSignUpResponse

Create a value of AdminConfirmSignUpResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:AdminConfirmSignUpResponse', adminConfirmSignUpResponse_httpStatus - The response's http status code.

Response Lenses

adminConfirmSignUpResponse_httpStatus :: Lens' AdminConfirmSignUpResponse Int

The response's http status code.