Network.AWS.CognitoIdentityProvider.AdminDisableProviderForUser

Description

Disables the user from signing in with the specified external (SAML or social) identity provider. If the user to disable is a Cognito User Pools native username + password user, they are not permitted to use their password to sign-in. If the user to disable is a linked external IdP user, any link between that user and an existing user is removed. The next time the external user (no longer attached to the previously linked DestinationUser) signs in, they must create a new user account. See AdminLinkProviderForUser.

This action is enabled only for admin access and requires developer credentials.

The ProviderName must match the value specified when creating an IdP for the pool.

To disable a native username + password user, the ProviderName value must be Cognito and the ProviderAttributeName must be Cognito_Subject, with the ProviderAttributeValue being the name that is used in the user pool for the user.

The ProviderAttributeName must always be Cognito_Subject for social identity providers. The ProviderAttributeValue must always be the exact subject that was used when the user was originally linked as a source user.

For de-linking a SAML identity, there are two scenarios. If the linked identity has not yet been used to sign-in, the ProviderAttributeName and ProviderAttributeValue must be the same values that were used for the SourceUser when the identities were originally linked using AdminLinkProviderForUser call. (If the linking was done with ProviderAttributeName set to Cognito_Subject, the same applies here). However, if the user has already signed in, the ProviderAttributeName must be Cognito_Subject and ProviderAttributeValue must be the subject of the SAML assertion.

Creating a Request

data AdminDisableProviderForUser

See: newAdminDisableProviderForUser smart constructor.

Constructors

AdminDisableProviderForUser'

Fields

userPoolId :: Text

The user pool ID for the user pool.

user :: ProviderUserIdentifierType

The user to be disabled.

newAdminDisableProviderForUser

Arguments

:: Text
$sel:userPoolId:AdminDisableProviderForUser'

-> ProviderUserIdentifierType
$sel:user:AdminDisableProviderForUser'

-> AdminDisableProviderForUser

Create a value of AdminDisableProviderForUser with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userPoolId:AdminDisableProviderForUser', adminDisableProviderForUser_userPoolId - The user pool ID for the user pool.

$sel:user:AdminDisableProviderForUser', adminDisableProviderForUser_user - The user to be disabled.

Request Lenses

adminDisableProviderForUser_userPoolId :: Lens' AdminDisableProviderForUser Text

The user pool ID for the user pool.

adminDisableProviderForUser_user :: Lens' AdminDisableProviderForUser ProviderUserIdentifierType

The user to be disabled.

Destructuring the Response

data AdminDisableProviderForUserResponse

See: newAdminDisableProviderForUserResponse smart constructor.

Constructors

AdminDisableProviderForUserResponse'

Fields

httpStatus :: Int

The response's http status code.

newAdminDisableProviderForUserResponse

Arguments

:: Int
$sel:httpStatus:AdminDisableProviderForUserResponse'

-> AdminDisableProviderForUserResponse

Create a value of AdminDisableProviderForUserResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:AdminDisableProviderForUserResponse', adminDisableProviderForUserResponse_httpStatus - The response's http status code.

Response Lenses

adminDisableProviderForUserResponse_httpStatus :: Lens' AdminDisableProviderForUserResponse Int

The response's http status code.