Network.AWS.CognitoIdentityProvider.AdminGetDevice

Description

Gets the device, as an administrator.

Calling this action requires developer credentials.

Creating a Request

data AdminGetDevice

Represents the request to get the device, as an administrator.

See: newAdminGetDevice smart constructor.

Constructors

AdminGetDevice'

Fields

deviceKey :: Text

The device key.

userPoolId :: Text

The user pool ID.

username :: Sensitive Text

The user name.

newAdminGetDevice

Arguments

:: Text
$sel:deviceKey:AdminGetDevice'

-> Text
$sel:userPoolId:AdminGetDevice'

-> Text
$sel:username:AdminGetDevice'

-> AdminGetDevice

Create a value of AdminGetDevice with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceKey:AdminGetDevice', adminGetDevice_deviceKey - The device key.

$sel:userPoolId:AdminGetDevice', adminGetDevice_userPoolId - The user pool ID.

$sel:username:AdminGetDevice', adminGetDevice_username - The user name.

Request Lenses

adminGetDevice_deviceKey :: Lens' AdminGetDevice Text

The device key.

adminGetDevice_userPoolId :: Lens' AdminGetDevice Text

The user pool ID.

adminGetDevice_username :: Lens' AdminGetDevice Text

The user name.

Destructuring the Response

data AdminGetDeviceResponse

Gets the device response, as an administrator.

See: newAdminGetDeviceResponse smart constructor.

Constructors

AdminGetDeviceResponse'

Fields

httpStatus :: Int

The response's http status code.

device :: DeviceType

The device.

newAdminGetDeviceResponse

Arguments

:: Int
$sel:httpStatus:AdminGetDeviceResponse'

-> DeviceType
$sel:device:AdminGetDeviceResponse'

-> AdminGetDeviceResponse

Create a value of AdminGetDeviceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:AdminGetDeviceResponse', adminGetDeviceResponse_httpStatus - The response's http status code.

$sel:device:AdminGetDeviceResponse', adminGetDeviceResponse_device - The device.

Response Lenses

adminGetDeviceResponse_httpStatus :: Lens' AdminGetDeviceResponse Int

The response's http status code.

adminGetDeviceResponse_device :: Lens' AdminGetDeviceResponse DeviceType

The device.