Network.AWS.CognitoIdentityProvider.AdminGetUser

Description

Gets the specified user by user name in a user pool as an administrator. Works on any user.

Calling this action requires developer credentials.

Creating a Request

data AdminGetUser

Represents the request to get the specified user as an administrator.

See: newAdminGetUser smart constructor.

Constructors

AdminGetUser'

Fields

userPoolId :: Text

The user pool ID for the user pool where you want to get information about the user.

username :: Sensitive Text

The user name of the user you wish to retrieve.

newAdminGetUser

Arguments

:: Text
$sel:userPoolId:AdminGetUser'

-> Text
$sel:username:AdminGetUser'

-> AdminGetUser

Create a value of AdminGetUser with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userPoolId:AdminGetUser', adminGetUser_userPoolId - The user pool ID for the user pool where you want to get information about the user.

$sel:username:AdminGetUser', adminGetUser_username - The user name of the user you wish to retrieve.

Request Lenses

adminGetUser_userPoolId :: Lens' AdminGetUser Text

The user pool ID for the user pool where you want to get information about the user.

adminGetUser_username :: Lens' AdminGetUser Text

The user name of the user you wish to retrieve.

Destructuring the Response

data AdminGetUserResponse

Represents the response from the server from the request to get the specified user as an administrator.

See: newAdminGetUserResponse smart constructor.

Constructors

AdminGetUserResponse'

Fields

enabled :: Maybe Bool

Indicates that the status is enabled.

userStatus :: Maybe UserStatusType

The user status. Can be one of the following:

UNCONFIRMED - User has been created but not confirmed.

CONFIRMED - User has been confirmed.

ARCHIVED - User is no longer active.

COMPROMISED - User is disabled due to a potential security threat.

UNKNOWN - User status is not known.

RESET_REQUIRED - User is confirmed, but the user must request a code and reset his or her password before he or she can sign in.

FORCE_CHANGE_PASSWORD - The user is confirmed and the user can sign in using a temporary password, but on first sign-in, the user must change his or her password to a new value before doing anything else.

userAttributes :: Maybe [AttributeType]

An array of name-value pairs representing user attributes.

userCreateDate :: Maybe POSIX

The date the user was created.

userMFASettingList :: Maybe [Text]

The MFA options that are enabled for the user. The possible values in this list are SMS_MFA and SOFTWARE_TOKEN_MFA.

mfaOptions :: Maybe [MFAOptionType]

This response parameter is no longer supported. It provides information only about SMS MFA configurations. It doesn't provide information about TOTP software token MFA configurations. To look up information about either type of MFA configuration, use UserMFASettingList instead.

userLastModifiedDate :: Maybe POSIX

The date the user was last modified.

preferredMfaSetting :: Maybe Text

The user's preferred MFA setting.

httpStatus :: Int

The response's http status code.

username :: Sensitive Text

The user name of the user about whom you are receiving information.

newAdminGetUserResponse

Arguments

:: Int
$sel:httpStatus:AdminGetUserResponse'

-> Text
$sel:username:AdminGetUser'

-> AdminGetUserResponse

Create a value of AdminGetUserResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:enabled:AdminGetUserResponse', adminGetUserResponse_enabled - Indicates that the status is enabled.

$sel:userStatus:AdminGetUserResponse', adminGetUserResponse_userStatus - The user status. Can be one of the following:

UNCONFIRMED - User has been created but not confirmed.

CONFIRMED - User has been confirmed.

ARCHIVED - User is no longer active.

COMPROMISED - User is disabled due to a potential security threat.

UNKNOWN - User status is not known.

RESET_REQUIRED - User is confirmed, but the user must request a code and reset his or her password before he or she can sign in.

FORCE_CHANGE_PASSWORD - The user is confirmed and the user can sign in using a temporary password, but on first sign-in, the user must change his or her password to a new value before doing anything else.

$sel:userAttributes:AdminGetUserResponse', adminGetUserResponse_userAttributes - An array of name-value pairs representing user attributes.

$sel:userCreateDate:AdminGetUserResponse', adminGetUserResponse_userCreateDate - The date the user was created.

$sel:userMFASettingList:AdminGetUserResponse', adminGetUserResponse_userMFASettingList - The MFA options that are enabled for the user. The possible values in this list are SMS_MFA and SOFTWARE_TOKEN_MFA.

$sel:mfaOptions:AdminGetUserResponse', adminGetUserResponse_mfaOptions - This response parameter is no longer supported. It provides information only about SMS MFA configurations. It doesn't provide information about TOTP software token MFA configurations. To look up information about either type of MFA configuration, use UserMFASettingList instead.

$sel:userLastModifiedDate:AdminGetUserResponse', adminGetUserResponse_userLastModifiedDate - The date the user was last modified.

$sel:preferredMfaSetting:AdminGetUserResponse', adminGetUserResponse_preferredMfaSetting - The user's preferred MFA setting.

$sel:httpStatus:AdminGetUserResponse', adminGetUserResponse_httpStatus - The response's http status code.

$sel:username:AdminGetUser', adminGetUserResponse_username - The user name of the user about whom you are receiving information.

Response Lenses

adminGetUserResponse_enabled :: Lens' AdminGetUserResponse (Maybe Bool)

Indicates that the status is enabled.

adminGetUserResponse_userStatus :: Lens' AdminGetUserResponse (Maybe UserStatusType)

The user status. Can be one of the following:

UNCONFIRMED - User has been created but not confirmed.

CONFIRMED - User has been confirmed.

ARCHIVED - User is no longer active.

COMPROMISED - User is disabled due to a potential security threat.

UNKNOWN - User status is not known.

RESET_REQUIRED - User is confirmed, but the user must request a code and reset his or her password before he or she can sign in.

FORCE_CHANGE_PASSWORD - The user is confirmed and the user can sign in using a temporary password, but on first sign-in, the user must change his or her password to a new value before doing anything else.

adminGetUserResponse_userAttributes :: Lens' AdminGetUserResponse (Maybe [AttributeType])

An array of name-value pairs representing user attributes.

adminGetUserResponse_userCreateDate :: Lens' AdminGetUserResponse (Maybe UTCTime)

The date the user was created.

adminGetUserResponse_userMFASettingList :: Lens' AdminGetUserResponse (Maybe [Text])

The MFA options that are enabled for the user. The possible values in this list are SMS_MFA and SOFTWARE_TOKEN_MFA.

adminGetUserResponse_mfaOptions :: Lens' AdminGetUserResponse (Maybe [MFAOptionType])

This response parameter is no longer supported. It provides information only about SMS MFA configurations. It doesn't provide information about TOTP software token MFA configurations. To look up information about either type of MFA configuration, use UserMFASettingList instead.

adminGetUserResponse_userLastModifiedDate :: Lens' AdminGetUserResponse (Maybe UTCTime)

The date the user was last modified.

adminGetUserResponse_preferredMfaSetting :: Lens' AdminGetUserResponse (Maybe Text)

The user's preferred MFA setting.

adminGetUserResponse_httpStatus :: Lens' AdminGetUserResponse Int

The response's http status code.

adminGetUserResponse_username :: Lens' AdminGetUserResponse Text

The user name of the user about whom you are receiving information.