Network.AWS.CognitoIdentityProvider.AdminResetUserPassword

Description

Resets the specified user's password in a user pool as an administrator. Works on any user.

When a developer calls this API, the current password is invalidated, so it must be changed. If a user tries to sign in after the API is called, the app will get a PasswordResetRequiredException exception back and should direct the user down the flow to reset the password, which is the same as the forgot password flow. In addition, if the user pool has phone verification selected and a verified phone number exists for the user, or if email verification is selected and a verified email exists for the user, calling this API will also result in sending a message to the end user with the code to change their password.

This action might generate an SMS text message. Starting June 1, 2021, U.S. telecom carriers require that you register an origination phone number before you can send SMS messages to U.S. phone numbers. If you use SMS text messages in Amazon Cognito, you must register a phone number with Amazon Pinpoint. Cognito will use the the registered number automatically. Otherwise, Cognito users that must receive SMS messages might be unable to sign up, activate their accounts, or sign in.

If you have never used SMS text messages with Amazon Cognito or any other Amazon Web Service, Amazon SNS might place your account in SMS sandbox. In <https:docs.aws.amazon.comsnslatestdgsns-sms-sandbox.html sandbox mode> , you’ll have limitations, such as sending messages to only verified phone numbers. After testing in the sandbox environment, you can move out of the SMS sandbox and into production. For more information, see SMS message settings for Cognito User Pools in the Amazon Cognito Developer Guide.

Calling this action requires developer credentials.

Creating a Request

data AdminResetUserPassword

Represents the request to reset a user's password as an administrator.

See: newAdminResetUserPassword smart constructor.

Constructors

AdminResetUserPassword'

Fields

clientMetadata :: Maybe (HashMap Text Text)

A map of custom key-value pairs that you can provide as input for any custom workflows that this action triggers.

You create custom workflows by assigning Lambda functions to user pool triggers. When you use the AdminResetUserPassword API action, Amazon Cognito invokes the function that is assigned to the custom message trigger. When Amazon Cognito invokes this function, it passes a JSON payload, which the function receives as input. This payload contains a clientMetadata attribute, which provides the data that you assigned to the ClientMetadata parameter in your AdminResetUserPassword request. In your function code in Lambda, you can process the clientMetadata value to enhance your workflow for your specific needs.

For more information, see Customizing User Pool Workflows with Lambda Triggers in the Amazon Cognito Developer Guide.

Take the following limitations into consideration when you use the ClientMetadata parameter:

Amazon Cognito does not store the ClientMetadata value. This data is available only to Lambda triggers that are assigned to a user pool to support custom workflows. If your user pool configuration does not include triggers, the ClientMetadata parameter serves no purpose.

Amazon Cognito does not validate the ClientMetadata value.

Amazon Cognito does not encrypt the the ClientMetadata value, so don't use it to provide sensitive information.

userPoolId :: Text

The user pool ID for the user pool where you want to reset the user's password.

username :: Sensitive Text

The user name of the user whose password you wish to reset.

newAdminResetUserPassword

Arguments

:: Text
$sel:userPoolId:AdminResetUserPassword'

-> Text
$sel:username:AdminResetUserPassword'

-> AdminResetUserPassword

Create a value of AdminResetUserPassword with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientMetadata:AdminResetUserPassword', adminResetUserPassword_clientMetadata - A map of custom key-value pairs that you can provide as input for any custom workflows that this action triggers.

You create custom workflows by assigning Lambda functions to user pool triggers. When you use the AdminResetUserPassword API action, Amazon Cognito invokes the function that is assigned to the custom message trigger. When Amazon Cognito invokes this function, it passes a JSON payload, which the function receives as input. This payload contains a clientMetadata attribute, which provides the data that you assigned to the ClientMetadata parameter in your AdminResetUserPassword request. In your function code in Lambda, you can process the clientMetadata value to enhance your workflow for your specific needs.

For more information, see Customizing User Pool Workflows with Lambda Triggers in the Amazon Cognito Developer Guide.

Take the following limitations into consideration when you use the ClientMetadata parameter:

Amazon Cognito does not store the ClientMetadata value. This data is available only to Lambda triggers that are assigned to a user pool to support custom workflows. If your user pool configuration does not include triggers, the ClientMetadata parameter serves no purpose.

Amazon Cognito does not validate the ClientMetadata value.

Amazon Cognito does not encrypt the the ClientMetadata value, so don't use it to provide sensitive information.

$sel:userPoolId:AdminResetUserPassword', adminResetUserPassword_userPoolId - The user pool ID for the user pool where you want to reset the user's password.

$sel:username:AdminResetUserPassword', adminResetUserPassword_username - The user name of the user whose password you wish to reset.

Request Lenses

adminResetUserPassword_clientMetadata :: Lens' AdminResetUserPassword (Maybe (HashMap Text Text))

A map of custom key-value pairs that you can provide as input for any custom workflows that this action triggers.

You create custom workflows by assigning Lambda functions to user pool triggers. When you use the AdminResetUserPassword API action, Amazon Cognito invokes the function that is assigned to the custom message trigger. When Amazon Cognito invokes this function, it passes a JSON payload, which the function receives as input. This payload contains a clientMetadata attribute, which provides the data that you assigned to the ClientMetadata parameter in your AdminResetUserPassword request. In your function code in Lambda, you can process the clientMetadata value to enhance your workflow for your specific needs.

For more information, see Customizing User Pool Workflows with Lambda Triggers in the Amazon Cognito Developer Guide.

Take the following limitations into consideration when you use the ClientMetadata parameter:

Amazon Cognito does not store the ClientMetadata value. This data is available only to Lambda triggers that are assigned to a user pool to support custom workflows. If your user pool configuration does not include triggers, the ClientMetadata parameter serves no purpose.

Amazon Cognito does not validate the ClientMetadata value.

Amazon Cognito does not encrypt the the ClientMetadata value, so don't use it to provide sensitive information.

adminResetUserPassword_userPoolId :: Lens' AdminResetUserPassword Text

The user pool ID for the user pool where you want to reset the user's password.

adminResetUserPassword_username :: Lens' AdminResetUserPassword Text

The user name of the user whose password you wish to reset.

Destructuring the Response

data AdminResetUserPasswordResponse

Represents the response from the server to reset a user password as an administrator.

See: newAdminResetUserPasswordResponse smart constructor.

Constructors

AdminResetUserPasswordResponse'

Fields

httpStatus :: Int

The response's http status code.

newAdminResetUserPasswordResponse

Arguments

:: Int
$sel:httpStatus:AdminResetUserPasswordResponse'

-> AdminResetUserPasswordResponse

Create a value of AdminResetUserPasswordResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:AdminResetUserPasswordResponse', adminResetUserPasswordResponse_httpStatus - The response's http status code.

Response Lenses

adminResetUserPasswordResponse_httpStatus :: Lens' AdminResetUserPasswordResponse Int

The response's http status code.