Network.AWS.CognitoIdentityProvider.AdminRespondToAuthChallenge

Description

Responds to an authentication challenge, as an administrator.

This action might generate an SMS text message. Starting June 1, 2021, U.S. telecom carriers require that you register an origination phone number before you can send SMS messages to U.S. phone numbers. If you use SMS text messages in Amazon Cognito, you must register a phone number with Amazon Pinpoint. Cognito will use the the registered number automatically. Otherwise, Cognito users that must receive SMS messages might be unable to sign up, activate their accounts, or sign in.

If you have never used SMS text messages with Amazon Cognito or any other Amazon Web Service, Amazon SNS might place your account in SMS sandbox. In <https:docs.aws.amazon.comsnslatestdgsns-sms-sandbox.html sandbox mode> , you’ll have limitations, such as sending messages to only verified phone numbers. After testing in the sandbox environment, you can move out of the SMS sandbox and into production. For more information, see SMS message settings for Cognito User Pools in the Amazon Cognito Developer Guide.

Calling this action requires developer credentials.

Creating a Request

data AdminRespondToAuthChallenge

The request to respond to the authentication challenge, as an administrator.

See: newAdminRespondToAuthChallenge smart constructor.

Constructors

AdminRespondToAuthChallenge'

Fields

clientMetadata :: Maybe (HashMap Text Text)

A map of custom key-value pairs that you can provide as input for any custom workflows that this action triggers.

You create custom workflows by assigning Lambda functions to user pool triggers. When you use the AdminRespondToAuthChallenge API action, Amazon Cognito invokes any functions that are assigned to the following triggers: pre sign-up, custom message, post authentication, /user migration, pre token generation, define auth challenge, create auth challenge, and verify auth challenge response/. When Amazon Cognito invokes any of these functions, it passes a JSON payload, which the function receives as input. This payload contains a clientMetadata attribute, which provides the data that you assigned to the ClientMetadata parameter in your AdminRespondToAuthChallenge request. In your function code in Lambda, you can process the clientMetadata value to enhance your workflow for your specific needs.

For more information, see Customizing User Pool Workflows with Lambda Triggers in the Amazon Cognito Developer Guide.

Take the following limitations into consideration when you use the ClientMetadata parameter:

Amazon Cognito does not store the ClientMetadata value. This data is available only to Lambda triggers that are assigned to a user pool to support custom workflows. If your user pool configuration does not include triggers, the ClientMetadata parameter serves no purpose.

Amazon Cognito does not validate the ClientMetadata value.

Amazon Cognito does not encrypt the the ClientMetadata value, so don't use it to provide sensitive information.

challengeResponses :: Maybe (HashMap Text Text)

The challenge responses. These are inputs corresponding to the value of ChallengeName, for example:

SMS_MFA: SMS_MFA_CODE, USERNAME, SECRET_HASH (if app client is configured with client secret).

PASSWORD_VERIFIER: PASSWORD_CLAIM_SIGNATURE, PASSWORD_CLAIM_SECRET_BLOCK, TIMESTAMP, USERNAME, SECRET_HASH (if app client is configured with client secret).

ADMIN_NO_SRP_AUTH: PASSWORD, USERNAME, SECRET_HASH (if app client is configured with client secret).

NEW_PASSWORD_REQUIRED: NEW_PASSWORD, any other required attributes, USERNAME, SECRET_HASH (if app client is configured with client secret).

MFA_SETUP requires USERNAME, plus you need to use the session value returned by VerifySoftwareToken in the Session parameter.

The value of the USERNAME attribute must be the user's actual username, not an alias (such as email address or phone number). To make this easier, the AdminInitiateAuth response includes the actual username value in the USERNAMEUSER_ID_FOR_SRP attribute, even if you specified an alias in your call to AdminInitiateAuth.

contextData :: Maybe ContextDataType

Contextual data such as the user's device fingerprint, IP address, or location used for evaluating the risk of an unexpected event by Amazon Cognito advanced security.

session :: Maybe Text

The session which should be passed both ways in challenge-response calls to the service. If InitiateAuth or RespondToAuthChallenge API call determines that the caller needs to go through another challenge, they return a session with other challenge parameters. This session should be passed as it is to the next RespondToAuthChallenge API call.

analyticsMetadata :: Maybe AnalyticsMetadataType

The analytics metadata for collecting Amazon Pinpoint metrics for AdminRespondToAuthChallenge calls.

userPoolId :: Text

The ID of the Amazon Cognito user pool.

clientId :: Sensitive Text

The app client ID.

challengeName :: ChallengeNameType

The challenge name. For more information, see AdminInitiateAuth.

newAdminRespondToAuthChallenge

Arguments

:: Text
$sel:userPoolId:AdminRespondToAuthChallenge'

-> Text
$sel:clientId:AdminRespondToAuthChallenge'

-> ChallengeNameType
$sel:challengeName:AdminRespondToAuthChallenge'

-> AdminRespondToAuthChallenge

Create a value of AdminRespondToAuthChallenge with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientMetadata:AdminRespondToAuthChallenge', adminRespondToAuthChallenge_clientMetadata - A map of custom key-value pairs that you can provide as input for any custom workflows that this action triggers.

You create custom workflows by assigning Lambda functions to user pool triggers. When you use the AdminRespondToAuthChallenge API action, Amazon Cognito invokes any functions that are assigned to the following triggers: pre sign-up, custom message, post authentication, /user migration, pre token generation, define auth challenge, create auth challenge, and verify auth challenge response/. When Amazon Cognito invokes any of these functions, it passes a JSON payload, which the function receives as input. This payload contains a clientMetadata attribute, which provides the data that you assigned to the ClientMetadata parameter in your AdminRespondToAuthChallenge request. In your function code in Lambda, you can process the clientMetadata value to enhance your workflow for your specific needs.

For more information, see Customizing User Pool Workflows with Lambda Triggers in the Amazon Cognito Developer Guide.

Take the following limitations into consideration when you use the ClientMetadata parameter:

Amazon Cognito does not store the ClientMetadata value. This data is available only to Lambda triggers that are assigned to a user pool to support custom workflows. If your user pool configuration does not include triggers, the ClientMetadata parameter serves no purpose.

Amazon Cognito does not validate the ClientMetadata value.

Amazon Cognito does not encrypt the the ClientMetadata value, so don't use it to provide sensitive information.

$sel:challengeResponses:AdminRespondToAuthChallenge', adminRespondToAuthChallenge_challengeResponses - The challenge responses. These are inputs corresponding to the value of ChallengeName, for example:

SMS_MFA: SMS_MFA_CODE, USERNAME, SECRET_HASH (if app client is configured with client secret).

PASSWORD_VERIFIER: PASSWORD_CLAIM_SIGNATURE, PASSWORD_CLAIM_SECRET_BLOCK, TIMESTAMP, USERNAME, SECRET_HASH (if app client is configured with client secret).

ADMIN_NO_SRP_AUTH: PASSWORD, USERNAME, SECRET_HASH (if app client is configured with client secret).

NEW_PASSWORD_REQUIRED: NEW_PASSWORD, any other required attributes, USERNAME, SECRET_HASH (if app client is configured with client secret).

MFA_SETUP requires USERNAME, plus you need to use the session value returned by VerifySoftwareToken in the Session parameter.

The value of the USERNAME attribute must be the user's actual username, not an alias (such as email address or phone number). To make this easier, the AdminInitiateAuth response includes the actual username value in the USERNAMEUSER_ID_FOR_SRP attribute, even if you specified an alias in your call to AdminInitiateAuth.

$sel:contextData:AdminRespondToAuthChallenge', adminRespondToAuthChallenge_contextData - Contextual data such as the user's device fingerprint, IP address, or location used for evaluating the risk of an unexpected event by Amazon Cognito advanced security.

$sel:session:AdminRespondToAuthChallenge', adminRespondToAuthChallenge_session - The session which should be passed both ways in challenge-response calls to the service. If InitiateAuth or RespondToAuthChallenge API call determines that the caller needs to go through another challenge, they return a session with other challenge parameters. This session should be passed as it is to the next RespondToAuthChallenge API call.

$sel:analyticsMetadata:AdminRespondToAuthChallenge', adminRespondToAuthChallenge_analyticsMetadata - The analytics metadata for collecting Amazon Pinpoint metrics for AdminRespondToAuthChallenge calls.

$sel:userPoolId:AdminRespondToAuthChallenge', adminRespondToAuthChallenge_userPoolId - The ID of the Amazon Cognito user pool.

$sel:clientId:AdminRespondToAuthChallenge', adminRespondToAuthChallenge_clientId - The app client ID.

$sel:challengeName:AdminRespondToAuthChallenge', adminRespondToAuthChallenge_challengeName - The challenge name. For more information, see AdminInitiateAuth.

Request Lenses

adminRespondToAuthChallenge_clientMetadata :: Lens' AdminRespondToAuthChallenge (Maybe (HashMap Text Text))

A map of custom key-value pairs that you can provide as input for any custom workflows that this action triggers.

You create custom workflows by assigning Lambda functions to user pool triggers. When you use the AdminRespondToAuthChallenge API action, Amazon Cognito invokes any functions that are assigned to the following triggers: pre sign-up, custom message, post authentication, /user migration, pre token generation, define auth challenge, create auth challenge, and verify auth challenge response/. When Amazon Cognito invokes any of these functions, it passes a JSON payload, which the function receives as input. This payload contains a clientMetadata attribute, which provides the data that you assigned to the ClientMetadata parameter in your AdminRespondToAuthChallenge request. In your function code in Lambda, you can process the clientMetadata value to enhance your workflow for your specific needs.

For more information, see Customizing User Pool Workflows with Lambda Triggers in the Amazon Cognito Developer Guide.

Take the following limitations into consideration when you use the ClientMetadata parameter:

Amazon Cognito does not store the ClientMetadata value. This data is available only to Lambda triggers that are assigned to a user pool to support custom workflows. If your user pool configuration does not include triggers, the ClientMetadata parameter serves no purpose.

Amazon Cognito does not validate the ClientMetadata value.

Amazon Cognito does not encrypt the the ClientMetadata value, so don't use it to provide sensitive information.

adminRespondToAuthChallenge_challengeResponses :: Lens' AdminRespondToAuthChallenge (Maybe (HashMap Text Text))

The challenge responses. These are inputs corresponding to the value of ChallengeName, for example:

SMS_MFA: SMS_MFA_CODE, USERNAME, SECRET_HASH (if app client is configured with client secret).

PASSWORD_VERIFIER: PASSWORD_CLAIM_SIGNATURE, PASSWORD_CLAIM_SECRET_BLOCK, TIMESTAMP, USERNAME, SECRET_HASH (if app client is configured with client secret).

ADMIN_NO_SRP_AUTH: PASSWORD, USERNAME, SECRET_HASH (if app client is configured with client secret).

NEW_PASSWORD_REQUIRED: NEW_PASSWORD, any other required attributes, USERNAME, SECRET_HASH (if app client is configured with client secret).

MFA_SETUP requires USERNAME, plus you need to use the session value returned by VerifySoftwareToken in the Session parameter.

The value of the USERNAME attribute must be the user's actual username, not an alias (such as email address or phone number). To make this easier, the AdminInitiateAuth response includes the actual username value in the USERNAMEUSER_ID_FOR_SRP attribute, even if you specified an alias in your call to AdminInitiateAuth.

adminRespondToAuthChallenge_contextData :: Lens' AdminRespondToAuthChallenge (Maybe ContextDataType)

Contextual data such as the user's device fingerprint, IP address, or location used for evaluating the risk of an unexpected event by Amazon Cognito advanced security.

adminRespondToAuthChallenge_session :: Lens' AdminRespondToAuthChallenge (Maybe Text)

The session which should be passed both ways in challenge-response calls to the service. If InitiateAuth or RespondToAuthChallenge API call determines that the caller needs to go through another challenge, they return a session with other challenge parameters. This session should be passed as it is to the next RespondToAuthChallenge API call.

adminRespondToAuthChallenge_analyticsMetadata :: Lens' AdminRespondToAuthChallenge (Maybe AnalyticsMetadataType)

The analytics metadata for collecting Amazon Pinpoint metrics for AdminRespondToAuthChallenge calls.

adminRespondToAuthChallenge_userPoolId :: Lens' AdminRespondToAuthChallenge Text

The ID of the Amazon Cognito user pool.

adminRespondToAuthChallenge_clientId :: Lens' AdminRespondToAuthChallenge Text

The app client ID.

adminRespondToAuthChallenge_challengeName :: Lens' AdminRespondToAuthChallenge ChallengeNameType

The challenge name. For more information, see AdminInitiateAuth.

Destructuring the Response

data AdminRespondToAuthChallengeResponse

Responds to the authentication challenge, as an administrator.

See: newAdminRespondToAuthChallengeResponse smart constructor.

Constructors

AdminRespondToAuthChallengeResponse'

Fields

authenticationResult :: Maybe AuthenticationResultType

The result returned by the server in response to the authentication request.

challengeName :: Maybe ChallengeNameType

The name of the challenge. For more information, see AdminInitiateAuth.

challengeParameters :: Maybe (HashMap Text Text)

The challenge parameters. For more information, see AdminInitiateAuth.

session :: Maybe Text

The session which should be passed both ways in challenge-response calls to the service. If the caller needs to go through another challenge, they return a session with other challenge parameters. This session should be passed as it is to the next RespondToAuthChallenge API call.

httpStatus :: Int

The response's http status code.

newAdminRespondToAuthChallengeResponse

Arguments

:: Int
$sel:httpStatus:AdminRespondToAuthChallengeResponse'

-> AdminRespondToAuthChallengeResponse

Create a value of AdminRespondToAuthChallengeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:authenticationResult:AdminRespondToAuthChallengeResponse', adminRespondToAuthChallengeResponse_authenticationResult - The result returned by the server in response to the authentication request.

$sel:challengeName:AdminRespondToAuthChallenge', adminRespondToAuthChallengeResponse_challengeName - The name of the challenge. For more information, see AdminInitiateAuth.

$sel:challengeParameters:AdminRespondToAuthChallengeResponse', adminRespondToAuthChallengeResponse_challengeParameters - The challenge parameters. For more information, see AdminInitiateAuth.

$sel:session:AdminRespondToAuthChallenge', adminRespondToAuthChallengeResponse_session - The session which should be passed both ways in challenge-response calls to the service. If the caller needs to go through another challenge, they return a session with other challenge parameters. This session should be passed as it is to the next RespondToAuthChallenge API call.

$sel:httpStatus:AdminRespondToAuthChallengeResponse', adminRespondToAuthChallengeResponse_httpStatus - The response's http status code.

Response Lenses

adminRespondToAuthChallengeResponse_authenticationResult :: Lens' AdminRespondToAuthChallengeResponse (Maybe AuthenticationResultType)

The result returned by the server in response to the authentication request.

adminRespondToAuthChallengeResponse_challengeName :: Lens' AdminRespondToAuthChallengeResponse (Maybe ChallengeNameType)

The name of the challenge. For more information, see AdminInitiateAuth.

adminRespondToAuthChallengeResponse_challengeParameters :: Lens' AdminRespondToAuthChallengeResponse (Maybe (HashMap Text Text))

The challenge parameters. For more information, see AdminInitiateAuth.

adminRespondToAuthChallengeResponse_session :: Lens' AdminRespondToAuthChallengeResponse (Maybe Text)

The session which should be passed both ways in challenge-response calls to the service. If the caller needs to go through another challenge, they return a session with other challenge parameters. This session should be passed as it is to the next RespondToAuthChallenge API call.

adminRespondToAuthChallengeResponse_httpStatus :: Lens' AdminRespondToAuthChallengeResponse Int

The response's http status code.