Network.AWS.CognitoIdentityProvider.AdminSetUserPassword

Description

Sets the specified user's password in a user pool as an administrator. Works on any user.

The password can be temporary or permanent. If it is temporary, the user status will be placed into the FORCE_CHANGE_PASSWORD state. When the user next tries to sign in, the InitiateAuth/AdminInitiateAuth response will contain the NEW_PASSWORD_REQUIRED challenge. If the user does not sign in before it expires, the user will not be able to sign in and their password will need to be reset by an administrator.

Once the user has set a new password, or the password is permanent, the user status will be set to Confirmed.

Creating a Request

data AdminSetUserPassword

See: newAdminSetUserPassword smart constructor.

Constructors

AdminSetUserPassword'

Fields

permanent :: Maybe Bool

True if the password is permanent, False if it is temporary.

userPoolId :: Text

The user pool ID for the user pool where you want to set the user's password.

username :: Sensitive Text

The user name of the user whose password you wish to set.

password :: Sensitive Text

The password for the user.

newAdminSetUserPassword

Arguments

:: Text
$sel:userPoolId:AdminSetUserPassword'

-> Text
$sel:username:AdminSetUserPassword'

-> Text
$sel:password:AdminSetUserPassword'

-> AdminSetUserPassword

Create a value of AdminSetUserPassword with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:permanent:AdminSetUserPassword', adminSetUserPassword_permanent - True if the password is permanent, False if it is temporary.

$sel:userPoolId:AdminSetUserPassword', adminSetUserPassword_userPoolId - The user pool ID for the user pool where you want to set the user's password.

$sel:username:AdminSetUserPassword', adminSetUserPassword_username - The user name of the user whose password you wish to set.

$sel:password:AdminSetUserPassword', adminSetUserPassword_password - The password for the user.

Request Lenses

adminSetUserPassword_permanent :: Lens' AdminSetUserPassword (Maybe Bool)

True if the password is permanent, False if it is temporary.

adminSetUserPassword_userPoolId :: Lens' AdminSetUserPassword Text

The user pool ID for the user pool where you want to set the user's password.

adminSetUserPassword_username :: Lens' AdminSetUserPassword Text

The user name of the user whose password you wish to set.

adminSetUserPassword_password :: Lens' AdminSetUserPassword Text

The password for the user.

Destructuring the Response

data AdminSetUserPasswordResponse

See: newAdminSetUserPasswordResponse smart constructor.

Constructors

AdminSetUserPasswordResponse'

Fields

httpStatus :: Int

The response's http status code.

newAdminSetUserPasswordResponse

Arguments

:: Int
$sel:httpStatus:AdminSetUserPasswordResponse'

-> AdminSetUserPasswordResponse

Create a value of AdminSetUserPasswordResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:AdminSetUserPasswordResponse', adminSetUserPasswordResponse_httpStatus - The response's http status code.

Response Lenses

adminSetUserPasswordResponse_httpStatus :: Lens' AdminSetUserPasswordResponse Int

The response's http status code.