Network.AWS.CognitoIdentityProvider.AdminUpdateDeviceStatus

Description

Updates the device status as an administrator.

Calling this action requires developer credentials.

Creating a Request

data AdminUpdateDeviceStatus

The request to update the device status, as an administrator.

See: newAdminUpdateDeviceStatus smart constructor.

Constructors

AdminUpdateDeviceStatus'

Fields

deviceRememberedStatus :: Maybe DeviceRememberedStatusType

The status indicating whether a device has been remembered or not.

userPoolId :: Text

The user pool ID.

username :: Sensitive Text

The user name.

deviceKey :: Text

The device key.

newAdminUpdateDeviceStatus

Arguments

:: Text
$sel:userPoolId:AdminUpdateDeviceStatus'

-> Text
$sel:username:AdminUpdateDeviceStatus'

-> Text
$sel:deviceKey:AdminUpdateDeviceStatus'

-> AdminUpdateDeviceStatus

Create a value of AdminUpdateDeviceStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceRememberedStatus:AdminUpdateDeviceStatus', adminUpdateDeviceStatus_deviceRememberedStatus - The status indicating whether a device has been remembered or not.

$sel:userPoolId:AdminUpdateDeviceStatus', adminUpdateDeviceStatus_userPoolId - The user pool ID.

$sel:username:AdminUpdateDeviceStatus', adminUpdateDeviceStatus_username - The user name.

$sel:deviceKey:AdminUpdateDeviceStatus', adminUpdateDeviceStatus_deviceKey - The device key.

Request Lenses

adminUpdateDeviceStatus_deviceRememberedStatus :: Lens' AdminUpdateDeviceStatus (Maybe DeviceRememberedStatusType)

The status indicating whether a device has been remembered or not.

adminUpdateDeviceStatus_userPoolId :: Lens' AdminUpdateDeviceStatus Text

The user pool ID.

adminUpdateDeviceStatus_username :: Lens' AdminUpdateDeviceStatus Text

The user name.

adminUpdateDeviceStatus_deviceKey :: Lens' AdminUpdateDeviceStatus Text

The device key.

Destructuring the Response

data AdminUpdateDeviceStatusResponse

The status response from the request to update the device, as an administrator.

See: newAdminUpdateDeviceStatusResponse smart constructor.

Constructors

AdminUpdateDeviceStatusResponse'

Fields

httpStatus :: Int

The response's http status code.

newAdminUpdateDeviceStatusResponse

Arguments

:: Int
$sel:httpStatus:AdminUpdateDeviceStatusResponse'

-> AdminUpdateDeviceStatusResponse

Create a value of AdminUpdateDeviceStatusResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:AdminUpdateDeviceStatusResponse', adminUpdateDeviceStatusResponse_httpStatus - The response's http status code.

Response Lenses

adminUpdateDeviceStatusResponse_httpStatus :: Lens' AdminUpdateDeviceStatusResponse Int

The response's http status code.