Network.AWS.CognitoIdentityProvider.AssociateSoftwareToken

Description

Returns a unique generated shared secret key code for the user account. The request takes an access token or a session string, but not both.

Calling AssociateSoftwareToken immediately disassociates the existing software token from the user account. If the user doesn't subsequently verify the software token, their account is essentially set up to authenticate without MFA. If MFA config is set to Optional at the user pool level, the user can then login without MFA. However, if MFA is set to Required for the user pool, the user will be asked to setup a new software token MFA during sign in.

Creating a Request

data AssociateSoftwareToken

See: newAssociateSoftwareToken smart constructor.

Constructors

AssociateSoftwareToken'

Fields

accessToken :: Maybe (Sensitive Text)

The access token.

session :: Maybe Text

The session which should be passed both ways in challenge-response calls to the service. This allows authentication of the user as part of the MFA setup process.

newAssociateSoftwareToken :: AssociateSoftwareToken

Create a value of AssociateSoftwareToken with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accessToken:AssociateSoftwareToken', associateSoftwareToken_accessToken - The access token.

$sel:session:AssociateSoftwareToken', associateSoftwareToken_session - The session which should be passed both ways in challenge-response calls to the service. This allows authentication of the user as part of the MFA setup process.

Request Lenses

associateSoftwareToken_accessToken :: Lens' AssociateSoftwareToken (Maybe Text)

The access token.

associateSoftwareToken_session :: Lens' AssociateSoftwareToken (Maybe Text)

The session which should be passed both ways in challenge-response calls to the service. This allows authentication of the user as part of the MFA setup process.

Destructuring the Response

data AssociateSoftwareTokenResponse

See: newAssociateSoftwareTokenResponse smart constructor.

Constructors

AssociateSoftwareTokenResponse'

Fields

secretCode :: Maybe (Sensitive Text)

A unique generated shared secret code that is used in the TOTP algorithm to generate a one time code.

session :: Maybe Text

The session which should be passed both ways in challenge-response calls to the service. This allows authentication of the user as part of the MFA setup process.

httpStatus :: Int

The response's http status code.

newAssociateSoftwareTokenResponse

Arguments

:: Int
$sel:httpStatus:AssociateSoftwareTokenResponse'

-> AssociateSoftwareTokenResponse

Create a value of AssociateSoftwareTokenResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:secretCode:AssociateSoftwareTokenResponse', associateSoftwareTokenResponse_secretCode - A unique generated shared secret code that is used in the TOTP algorithm to generate a one time code.

$sel:session:AssociateSoftwareToken', associateSoftwareTokenResponse_session - The session which should be passed both ways in challenge-response calls to the service. This allows authentication of the user as part of the MFA setup process.

$sel:httpStatus:AssociateSoftwareTokenResponse', associateSoftwareTokenResponse_httpStatus - The response's http status code.

Response Lenses

associateSoftwareTokenResponse_secretCode :: Lens' AssociateSoftwareTokenResponse (Maybe Text)

A unique generated shared secret code that is used in the TOTP algorithm to generate a one time code.

associateSoftwareTokenResponse_session :: Lens' AssociateSoftwareTokenResponse (Maybe Text)

The session which should be passed both ways in challenge-response calls to the service. This allows authentication of the user as part of the MFA setup process.

associateSoftwareTokenResponse_httpStatus :: Lens' AssociateSoftwareTokenResponse Int

The response's http status code.