Network.AWS.CognitoIdentityProvider.ConfirmSignUp

Description

Confirms registration of a user and handles the existing alias from a previous user.

Creating a Request

data ConfirmSignUp

Represents the request to confirm registration of a user.

See: newConfirmSignUp smart constructor.

Constructors

ConfirmSignUp'

Fields

clientMetadata :: Maybe (HashMap Text Text)

A map of custom key-value pairs that you can provide as input for any custom workflows that this action triggers.

You create custom workflows by assigning Lambda functions to user pool triggers. When you use the ConfirmSignUp API action, Amazon Cognito invokes the function that is assigned to the post confirmation trigger. When Amazon Cognito invokes this function, it passes a JSON payload, which the function receives as input. This payload contains a clientMetadata attribute, which provides the data that you assigned to the ClientMetadata parameter in your ConfirmSignUp request. In your function code in Lambda, you can process the clientMetadata value to enhance your workflow for your specific needs.

For more information, see Customizing User Pool Workflows with Lambda Triggers in the Amazon Cognito Developer Guide.

Take the following limitations into consideration when you use the ClientMetadata parameter:

Amazon Cognito does not store the ClientMetadata value. This data is available only to Lambda triggers that are assigned to a user pool to support custom workflows. If your user pool configuration does not include triggers, the ClientMetadata parameter serves no purpose.

Amazon Cognito does not validate the ClientMetadata value.

Amazon Cognito does not encrypt the the ClientMetadata value, so don't use it to provide sensitive information.

userContextData :: Maybe UserContextDataType

Contextual data such as the user's device fingerprint, IP address, or location used for evaluating the risk of an unexpected event by Amazon Cognito advanced security.

forceAliasCreation :: Maybe Bool

Boolean to be specified to force user confirmation irrespective of existing alias. By default set to False. If this parameter is set to True and the phone number/email used for sign up confirmation already exists as an alias with a different user, the API call will migrate the alias from the previous user to the newly created user being confirmed. If set to False, the API will throw an AliasExistsException error.

secretHash :: Maybe (Sensitive Text)

A keyed-hash message authentication code (HMAC) calculated using the secret key of a user pool client and username plus the client ID in the message.

analyticsMetadata :: Maybe AnalyticsMetadataType

The Amazon Pinpoint analytics metadata for collecting metrics for ConfirmSignUp calls.

clientId :: Sensitive Text

The ID of the app client associated with the user pool.

username :: Sensitive Text

The user name of the user whose registration you wish to confirm.

confirmationCode :: Text

The confirmation code sent by a user's request to confirm registration.

newConfirmSignUp

Arguments

:: Text
$sel:clientId:ConfirmSignUp'

-> Text
$sel:username:ConfirmSignUp'

-> Text
$sel:confirmationCode:ConfirmSignUp'

-> ConfirmSignUp

Create a value of ConfirmSignUp with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientMetadata:ConfirmSignUp', confirmSignUp_clientMetadata - A map of custom key-value pairs that you can provide as input for any custom workflows that this action triggers.

You create custom workflows by assigning Lambda functions to user pool triggers. When you use the ConfirmSignUp API action, Amazon Cognito invokes the function that is assigned to the post confirmation trigger. When Amazon Cognito invokes this function, it passes a JSON payload, which the function receives as input. This payload contains a clientMetadata attribute, which provides the data that you assigned to the ClientMetadata parameter in your ConfirmSignUp request. In your function code in Lambda, you can process the clientMetadata value to enhance your workflow for your specific needs.

For more information, see Customizing User Pool Workflows with Lambda Triggers in the Amazon Cognito Developer Guide.

Take the following limitations into consideration when you use the ClientMetadata parameter:

Amazon Cognito does not store the ClientMetadata value. This data is available only to Lambda triggers that are assigned to a user pool to support custom workflows. If your user pool configuration does not include triggers, the ClientMetadata parameter serves no purpose.

Amazon Cognito does not validate the ClientMetadata value.

Amazon Cognito does not encrypt the the ClientMetadata value, so don't use it to provide sensitive information.

$sel:userContextData:ConfirmSignUp', confirmSignUp_userContextData - Contextual data such as the user's device fingerprint, IP address, or location used for evaluating the risk of an unexpected event by Amazon Cognito advanced security.

$sel:forceAliasCreation:ConfirmSignUp', confirmSignUp_forceAliasCreation - Boolean to be specified to force user confirmation irrespective of existing alias. By default set to False. If this parameter is set to True and the phone number/email used for sign up confirmation already exists as an alias with a different user, the API call will migrate the alias from the previous user to the newly created user being confirmed. If set to False, the API will throw an AliasExistsException error.

$sel:secretHash:ConfirmSignUp', confirmSignUp_secretHash - A keyed-hash message authentication code (HMAC) calculated using the secret key of a user pool client and username plus the client ID in the message.

$sel:analyticsMetadata:ConfirmSignUp', confirmSignUp_analyticsMetadata - The Amazon Pinpoint analytics metadata for collecting metrics for ConfirmSignUp calls.

$sel:clientId:ConfirmSignUp', confirmSignUp_clientId - The ID of the app client associated with the user pool.

$sel:username:ConfirmSignUp', confirmSignUp_username - The user name of the user whose registration you wish to confirm.

$sel:confirmationCode:ConfirmSignUp', confirmSignUp_confirmationCode - The confirmation code sent by a user's request to confirm registration.

Request Lenses

confirmSignUp_clientMetadata :: Lens' ConfirmSignUp (Maybe (HashMap Text Text))

A map of custom key-value pairs that you can provide as input for any custom workflows that this action triggers.

You create custom workflows by assigning Lambda functions to user pool triggers. When you use the ConfirmSignUp API action, Amazon Cognito invokes the function that is assigned to the post confirmation trigger. When Amazon Cognito invokes this function, it passes a JSON payload, which the function receives as input. This payload contains a clientMetadata attribute, which provides the data that you assigned to the ClientMetadata parameter in your ConfirmSignUp request. In your function code in Lambda, you can process the clientMetadata value to enhance your workflow for your specific needs.

For more information, see Customizing User Pool Workflows with Lambda Triggers in the Amazon Cognito Developer Guide.

Take the following limitations into consideration when you use the ClientMetadata parameter:

Amazon Cognito does not store the ClientMetadata value. This data is available only to Lambda triggers that are assigned to a user pool to support custom workflows. If your user pool configuration does not include triggers, the ClientMetadata parameter serves no purpose.

Amazon Cognito does not validate the ClientMetadata value.

Amazon Cognito does not encrypt the the ClientMetadata value, so don't use it to provide sensitive information.

confirmSignUp_userContextData :: Lens' ConfirmSignUp (Maybe UserContextDataType)

Contextual data such as the user's device fingerprint, IP address, or location used for evaluating the risk of an unexpected event by Amazon Cognito advanced security.

confirmSignUp_forceAliasCreation :: Lens' ConfirmSignUp (Maybe Bool)

Boolean to be specified to force user confirmation irrespective of existing alias. By default set to False. If this parameter is set to True and the phone number/email used for sign up confirmation already exists as an alias with a different user, the API call will migrate the alias from the previous user to the newly created user being confirmed. If set to False, the API will throw an AliasExistsException error.

confirmSignUp_secretHash :: Lens' ConfirmSignUp (Maybe Text)

A keyed-hash message authentication code (HMAC) calculated using the secret key of a user pool client and username plus the client ID in the message.

confirmSignUp_analyticsMetadata :: Lens' ConfirmSignUp (Maybe AnalyticsMetadataType)

The Amazon Pinpoint analytics metadata for collecting metrics for ConfirmSignUp calls.

confirmSignUp_clientId :: Lens' ConfirmSignUp Text

The ID of the app client associated with the user pool.

confirmSignUp_username :: Lens' ConfirmSignUp Text

The user name of the user whose registration you wish to confirm.

confirmSignUp_confirmationCode :: Lens' ConfirmSignUp Text

The confirmation code sent by a user's request to confirm registration.

Destructuring the Response

data ConfirmSignUpResponse

Represents the response from the server for the registration confirmation.

See: newConfirmSignUpResponse smart constructor.

Constructors

ConfirmSignUpResponse'

Fields

httpStatus :: Int

The response's http status code.

newConfirmSignUpResponse

Arguments

:: Int
$sel:httpStatus:ConfirmSignUpResponse'

-> ConfirmSignUpResponse

Create a value of ConfirmSignUpResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:ConfirmSignUpResponse', confirmSignUpResponse_httpStatus - The response's http status code.

Response Lenses

confirmSignUpResponse_httpStatus :: Lens' ConfirmSignUpResponse Int

The response's http status code.