Network.AWS.CognitoIdentityProvider.CreateUserPoolDomain

Description

Creates a new domain for a user pool.

Creating a Request

data CreateUserPoolDomain

See: newCreateUserPoolDomain smart constructor.

Constructors

CreateUserPoolDomain'

Fields

customDomainConfig :: Maybe CustomDomainConfigType

The configuration for a custom domain that hosts the sign-up and sign-in webpages for your application.

Provide this parameter only if you want to use a custom domain for your user pool. Otherwise, you can exclude this parameter and use the Amazon Cognito hosted domain instead.

For more information about the hosted domain and custom domains, see Configuring a User Pool Domain.

domain :: Text

The domain string.

userPoolId :: Text

The user pool ID.

newCreateUserPoolDomain

Arguments

:: Text
$sel:domain:CreateUserPoolDomain'

-> Text
$sel:userPoolId:CreateUserPoolDomain'

-> CreateUserPoolDomain

Create a value of CreateUserPoolDomain with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:customDomainConfig:CreateUserPoolDomain', createUserPoolDomain_customDomainConfig - The configuration for a custom domain that hosts the sign-up and sign-in webpages for your application.

Provide this parameter only if you want to use a custom domain for your user pool. Otherwise, you can exclude this parameter and use the Amazon Cognito hosted domain instead.

For more information about the hosted domain and custom domains, see Configuring a User Pool Domain.

$sel:domain:CreateUserPoolDomain', createUserPoolDomain_domain - The domain string.

$sel:userPoolId:CreateUserPoolDomain', createUserPoolDomain_userPoolId - The user pool ID.

Request Lenses

createUserPoolDomain_customDomainConfig :: Lens' CreateUserPoolDomain (Maybe CustomDomainConfigType)

The configuration for a custom domain that hosts the sign-up and sign-in webpages for your application.

Provide this parameter only if you want to use a custom domain for your user pool. Otherwise, you can exclude this parameter and use the Amazon Cognito hosted domain instead.

For more information about the hosted domain and custom domains, see Configuring a User Pool Domain.

createUserPoolDomain_domain :: Lens' CreateUserPoolDomain Text

The domain string.

createUserPoolDomain_userPoolId :: Lens' CreateUserPoolDomain Text

The user pool ID.

Destructuring the Response

data CreateUserPoolDomainResponse

See: newCreateUserPoolDomainResponse smart constructor.

Constructors

CreateUserPoolDomainResponse'

Fields

cloudFrontDomain :: Maybe Text

The Amazon CloudFront endpoint that you use as the target of the alias that you set up with your Domain Name Service (DNS) provider.

httpStatus :: Int

The response's http status code.

newCreateUserPoolDomainResponse

Arguments

:: Int
$sel:httpStatus:CreateUserPoolDomainResponse'

-> CreateUserPoolDomainResponse

Create a value of CreateUserPoolDomainResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cloudFrontDomain:CreateUserPoolDomainResponse', createUserPoolDomainResponse_cloudFrontDomain - The Amazon CloudFront endpoint that you use as the target of the alias that you set up with your Domain Name Service (DNS) provider.

$sel:httpStatus:CreateUserPoolDomainResponse', createUserPoolDomainResponse_httpStatus - The response's http status code.

Response Lenses

createUserPoolDomainResponse_cloudFrontDomain :: Lens' CreateUserPoolDomainResponse (Maybe Text)

The Amazon CloudFront endpoint that you use as the target of the alias that you set up with your Domain Name Service (DNS) provider.

createUserPoolDomainResponse_httpStatus :: Lens' CreateUserPoolDomainResponse Int

The response's http status code.