Network.AWS.CognitoIdentityProvider.DescribeUserImportJob

Description

Describes the user import job.

Creating a Request

data DescribeUserImportJob

Represents the request to describe the user import job.

See: newDescribeUserImportJob smart constructor.

Constructors

DescribeUserImportJob'

Fields

userPoolId :: Text

The user pool ID for the user pool that the users are being imported into.

jobId :: Text

The job ID for the user import job.

newDescribeUserImportJob

Arguments

:: Text
$sel:userPoolId:DescribeUserImportJob'

-> Text
$sel:jobId:DescribeUserImportJob'

-> DescribeUserImportJob

Create a value of DescribeUserImportJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userPoolId:DescribeUserImportJob', describeUserImportJob_userPoolId - The user pool ID for the user pool that the users are being imported into.

$sel:jobId:DescribeUserImportJob', describeUserImportJob_jobId - The job ID for the user import job.

Request Lenses

describeUserImportJob_userPoolId :: Lens' DescribeUserImportJob Text

The user pool ID for the user pool that the users are being imported into.

describeUserImportJob_jobId :: Lens' DescribeUserImportJob Text

The job ID for the user import job.

Destructuring the Response

data DescribeUserImportJobResponse

Represents the response from the server to the request to describe the user import job.

See: newDescribeUserImportJobResponse smart constructor.

Constructors

DescribeUserImportJobResponse'

Fields

userImportJob :: Maybe UserImportJobType

The job object that represents the user import job.

httpStatus :: Int

The response's http status code.

newDescribeUserImportJobResponse

Arguments

:: Int
$sel:httpStatus:DescribeUserImportJobResponse'

-> DescribeUserImportJobResponse

Create a value of DescribeUserImportJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userImportJob:DescribeUserImportJobResponse', describeUserImportJobResponse_userImportJob - The job object that represents the user import job.

$sel:httpStatus:DescribeUserImportJobResponse', describeUserImportJobResponse_httpStatus - The response's http status code.

Response Lenses

describeUserImportJobResponse_userImportJob :: Lens' DescribeUserImportJobResponse (Maybe UserImportJobType)

The job object that represents the user import job.

describeUserImportJobResponse_httpStatus :: Lens' DescribeUserImportJobResponse Int

The response's http status code.