Network.AWS.CognitoIdentityProvider.ForgotPassword

Description

Calling this API causes a message to be sent to the end user with a confirmation code that is required to change the user's password. For the Username parameter, you can use the username or user alias. The method used to send the confirmation code is sent according to the specified AccountRecoverySetting. For more information, see Recovering User Accounts in the Amazon Cognito Developer Guide. If neither a verified phone number nor a verified email exists, an InvalidParameterException is thrown. To use the confirmation code for resetting the password, call ConfirmForgotPassword.

This action might generate an SMS text message. Starting June 1, 2021, U.S. telecom carriers require that you register an origination phone number before you can send SMS messages to U.S. phone numbers. If you use SMS text messages in Amazon Cognito, you must register a phone number with Amazon Pinpoint. Cognito will use the the registered number automatically. Otherwise, Cognito users that must receive SMS messages might be unable to sign up, activate their accounts, or sign in.

If you have never used SMS text messages with Amazon Cognito or any other Amazon Web Service, Amazon SNS might place your account in SMS sandbox. In <https:docs.aws.amazon.comsnslatestdgsns-sms-sandbox.html sandbox mode> , you’ll have limitations, such as sending messages to only verified phone numbers. After testing in the sandbox environment, you can move out of the SMS sandbox and into production. For more information, see SMS message settings for Cognito User Pools in the Amazon Cognito Developer Guide.

Creating a Request

data ForgotPassword

Represents the request to reset a user's password.

See: newForgotPassword smart constructor.

Constructors

ForgotPassword'

Fields

clientMetadata :: Maybe (HashMap Text Text)

A map of custom key-value pairs that you can provide as input for any custom workflows that this action triggers.

You create custom workflows by assigning Lambda functions to user pool triggers. When you use the ForgotPassword API action, Amazon Cognito invokes any functions that are assigned to the following triggers: /pre sign-up, custom message, and user migration/. When Amazon Cognito invokes any of these functions, it passes a JSON payload, which the function receives as input. This payload contains a clientMetadata attribute, which provides the data that you assigned to the ClientMetadata parameter in your ForgotPassword request. In your function code in Lambda, you can process the clientMetadata value to enhance your workflow for your specific needs.

For more information, see Customizing User Pool Workflows with Lambda Triggers in the Amazon Cognito Developer Guide.

Take the following limitations into consideration when you use the ClientMetadata parameter:

Amazon Cognito does not store the ClientMetadata value. This data is available only to Lambda triggers that are assigned to a user pool to support custom workflows. If your user pool configuration does not include triggers, the ClientMetadata parameter serves no purpose.

Amazon Cognito does not validate the ClientMetadata value.

Amazon Cognito does not encrypt the the ClientMetadata value, so don't use it to provide sensitive information.

userContextData :: Maybe UserContextDataType

Contextual data such as the user's device fingerprint, IP address, or location used for evaluating the risk of an unexpected event by Amazon Cognito advanced security.

secretHash :: Maybe (Sensitive Text)

A keyed-hash message authentication code (HMAC) calculated using the secret key of a user pool client and username plus the client ID in the message.

analyticsMetadata :: Maybe AnalyticsMetadataType

The Amazon Pinpoint analytics metadata for collecting metrics for ForgotPassword calls.

clientId :: Sensitive Text

The ID of the client associated with the user pool.

username :: Sensitive Text

The user name of the user for whom you want to enter a code to reset a forgotten password.

newForgotPassword

Arguments

:: Text
$sel:clientId:ForgotPassword'

-> Text
$sel:username:ForgotPassword'

-> ForgotPassword

Create a value of ForgotPassword with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientMetadata:ForgotPassword', forgotPassword_clientMetadata - A map of custom key-value pairs that you can provide as input for any custom workflows that this action triggers.

You create custom workflows by assigning Lambda functions to user pool triggers. When you use the ForgotPassword API action, Amazon Cognito invokes any functions that are assigned to the following triggers: /pre sign-up, custom message, and user migration/. When Amazon Cognito invokes any of these functions, it passes a JSON payload, which the function receives as input. This payload contains a clientMetadata attribute, which provides the data that you assigned to the ClientMetadata parameter in your ForgotPassword request. In your function code in Lambda, you can process the clientMetadata value to enhance your workflow for your specific needs.

For more information, see Customizing User Pool Workflows with Lambda Triggers in the Amazon Cognito Developer Guide.

Take the following limitations into consideration when you use the ClientMetadata parameter:

Amazon Cognito does not store the ClientMetadata value. This data is available only to Lambda triggers that are assigned to a user pool to support custom workflows. If your user pool configuration does not include triggers, the ClientMetadata parameter serves no purpose.

Amazon Cognito does not validate the ClientMetadata value.

Amazon Cognito does not encrypt the the ClientMetadata value, so don't use it to provide sensitive information.

$sel:userContextData:ForgotPassword', forgotPassword_userContextData - Contextual data such as the user's device fingerprint, IP address, or location used for evaluating the risk of an unexpected event by Amazon Cognito advanced security.

$sel:secretHash:ForgotPassword', forgotPassword_secretHash - A keyed-hash message authentication code (HMAC) calculated using the secret key of a user pool client and username plus the client ID in the message.

$sel:analyticsMetadata:ForgotPassword', forgotPassword_analyticsMetadata - The Amazon Pinpoint analytics metadata for collecting metrics for ForgotPassword calls.

$sel:clientId:ForgotPassword', forgotPassword_clientId - The ID of the client associated with the user pool.

$sel:username:ForgotPassword', forgotPassword_username - The user name of the user for whom you want to enter a code to reset a forgotten password.

Request Lenses

forgotPassword_clientMetadata :: Lens' ForgotPassword (Maybe (HashMap Text Text))

A map of custom key-value pairs that you can provide as input for any custom workflows that this action triggers.

You create custom workflows by assigning Lambda functions to user pool triggers. When you use the ForgotPassword API action, Amazon Cognito invokes any functions that are assigned to the following triggers: /pre sign-up, custom message, and user migration/. When Amazon Cognito invokes any of these functions, it passes a JSON payload, which the function receives as input. This payload contains a clientMetadata attribute, which provides the data that you assigned to the ClientMetadata parameter in your ForgotPassword request. In your function code in Lambda, you can process the clientMetadata value to enhance your workflow for your specific needs.

For more information, see Customizing User Pool Workflows with Lambda Triggers in the Amazon Cognito Developer Guide.

Take the following limitations into consideration when you use the ClientMetadata parameter:

Amazon Cognito does not store the ClientMetadata value. This data is available only to Lambda triggers that are assigned to a user pool to support custom workflows. If your user pool configuration does not include triggers, the ClientMetadata parameter serves no purpose.

Amazon Cognito does not validate the ClientMetadata value.

Amazon Cognito does not encrypt the the ClientMetadata value, so don't use it to provide sensitive information.

forgotPassword_userContextData :: Lens' ForgotPassword (Maybe UserContextDataType)

Contextual data such as the user's device fingerprint, IP address, or location used for evaluating the risk of an unexpected event by Amazon Cognito advanced security.

forgotPassword_secretHash :: Lens' ForgotPassword (Maybe Text)

A keyed-hash message authentication code (HMAC) calculated using the secret key of a user pool client and username plus the client ID in the message.

forgotPassword_analyticsMetadata :: Lens' ForgotPassword (Maybe AnalyticsMetadataType)

The Amazon Pinpoint analytics metadata for collecting metrics for ForgotPassword calls.

forgotPassword_clientId :: Lens' ForgotPassword Text

The ID of the client associated with the user pool.

forgotPassword_username :: Lens' ForgotPassword Text

The user name of the user for whom you want to enter a code to reset a forgotten password.

Destructuring the Response

data ForgotPasswordResponse

Respresents the response from the server regarding the request to reset a password.

See: newForgotPasswordResponse smart constructor.

Constructors

ForgotPasswordResponse'

Fields

codeDeliveryDetails :: Maybe CodeDeliveryDetailsType

The code delivery details returned by the server in response to the request to reset a password.

httpStatus :: Int

The response's http status code.

newForgotPasswordResponse

Arguments

:: Int
$sel:httpStatus:ForgotPasswordResponse'

-> ForgotPasswordResponse

Create a value of ForgotPasswordResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:codeDeliveryDetails:ForgotPasswordResponse', forgotPasswordResponse_codeDeliveryDetails - The code delivery details returned by the server in response to the request to reset a password.

$sel:httpStatus:ForgotPasswordResponse', forgotPasswordResponse_httpStatus - The response's http status code.

Response Lenses

forgotPasswordResponse_codeDeliveryDetails :: Lens' ForgotPasswordResponse (Maybe CodeDeliveryDetailsType)

The code delivery details returned by the server in response to the request to reset a password.

forgotPasswordResponse_httpStatus :: Lens' ForgotPasswordResponse Int

The response's http status code.