Network.AWS.CognitoIdentityProvider.GetDevice

Description

Gets the device.

Creating a Request

data GetDevice

Represents the request to get the device.

See: newGetDevice smart constructor.

Constructors

GetDevice'

Fields

accessToken :: Maybe (Sensitive Text)

The access token.

deviceKey :: Text

The device key.

newGetDevice

Arguments

:: Text
$sel:deviceKey:GetDevice'

-> GetDevice

Create a value of GetDevice with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accessToken:GetDevice', getDevice_accessToken - The access token.

$sel:deviceKey:GetDevice', getDevice_deviceKey - The device key.

Request Lenses

getDevice_accessToken :: Lens' GetDevice (Maybe Text)

The access token.

getDevice_deviceKey :: Lens' GetDevice Text

The device key.

Destructuring the Response

data GetDeviceResponse

Gets the device response.

See: newGetDeviceResponse smart constructor.

Constructors

GetDeviceResponse'

Fields

httpStatus :: Int

The response's http status code.

device :: DeviceType

The device.

newGetDeviceResponse

Arguments

:: Int
$sel:httpStatus:GetDeviceResponse'

-> DeviceType
$sel:device:GetDeviceResponse'

-> GetDeviceResponse

Create a value of GetDeviceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetDeviceResponse', getDeviceResponse_httpStatus - The response's http status code.

$sel:device:GetDeviceResponse', getDeviceResponse_device - The device.

Response Lenses

getDeviceResponse_httpStatus :: Lens' GetDeviceResponse Int

The response's http status code.

getDeviceResponse_device :: Lens' GetDeviceResponse DeviceType

The device.