Network.AWS.CognitoIdentityProvider.GetUserAttributeVerificationCode

Description

Gets the user attribute verification code for the specified attribute name.

This action might generate an SMS text message. Starting June 1, 2021, U.S. telecom carriers require that you register an origination phone number before you can send SMS messages to U.S. phone numbers. If you use SMS text messages in Amazon Cognito, you must register a phone number with Amazon Pinpoint. Cognito will use the the registered number automatically. Otherwise, Cognito users that must receive SMS messages might be unable to sign up, activate their accounts, or sign in.

If you have never used SMS text messages with Amazon Cognito or any other Amazon Web Service, Amazon SNS might place your account in SMS sandbox. In <https:docs.aws.amazon.comsnslatestdgsns-sms-sandbox.html sandbox mode> , you’ll have limitations, such as sending messages to only verified phone numbers. After testing in the sandbox environment, you can move out of the SMS sandbox and into production. For more information, see SMS message settings for Cognito User Pools in the Amazon Cognito Developer Guide.

Creating a Request

data GetUserAttributeVerificationCode

Represents the request to get user attribute verification.

See: newGetUserAttributeVerificationCode smart constructor.

Constructors

GetUserAttributeVerificationCode'

Fields

clientMetadata :: Maybe (HashMap Text Text)

A map of custom key-value pairs that you can provide as input for any custom workflows that this action triggers.

You create custom workflows by assigning Lambda functions to user pool triggers. When you use the GetUserAttributeVerificationCode API action, Amazon Cognito invokes the function that is assigned to the /custom message/ trigger. When Amazon Cognito invokes this function, it passes a JSON payload, which the function receives as input. This payload contains a clientMetadata attribute, which provides the data that you assigned to the ClientMetadata parameter in your GetUserAttributeVerificationCode request. In your function code in Lambda, you can process the clientMetadata value to enhance your workflow for your specific needs.

For more information, see Customizing User Pool Workflows with Lambda Triggers in the Amazon Cognito Developer Guide.

Take the following limitations into consideration when you use the ClientMetadata parameter:

Amazon Cognito does not store the ClientMetadata value. This data is available only to Lambda triggers that are assigned to a user pool to support custom workflows. If your user pool configuration does not include triggers, the ClientMetadata parameter serves no purpose.

Amazon Cognito does not validate the ClientMetadata value.

Amazon Cognito does not encrypt the the ClientMetadata value, so don't use it to provide sensitive information.

accessToken :: Sensitive Text

The access token returned by the server response to get the user attribute verification code.

attributeName :: Text

The attribute name returned by the server response to get the user attribute verification code.

newGetUserAttributeVerificationCode

Arguments

:: Text
$sel:accessToken:GetUserAttributeVerificationCode'

-> Text
$sel:attributeName:GetUserAttributeVerificationCode'

-> GetUserAttributeVerificationCode

Create a value of GetUserAttributeVerificationCode with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientMetadata:GetUserAttributeVerificationCode', getUserAttributeVerificationCode_clientMetadata - A map of custom key-value pairs that you can provide as input for any custom workflows that this action triggers.

You create custom workflows by assigning Lambda functions to user pool triggers. When you use the GetUserAttributeVerificationCode API action, Amazon Cognito invokes the function that is assigned to the /custom message/ trigger. When Amazon Cognito invokes this function, it passes a JSON payload, which the function receives as input. This payload contains a clientMetadata attribute, which provides the data that you assigned to the ClientMetadata parameter in your GetUserAttributeVerificationCode request. In your function code in Lambda, you can process the clientMetadata value to enhance your workflow for your specific needs.

For more information, see Customizing User Pool Workflows with Lambda Triggers in the Amazon Cognito Developer Guide.

Take the following limitations into consideration when you use the ClientMetadata parameter:

Amazon Cognito does not store the ClientMetadata value. This data is available only to Lambda triggers that are assigned to a user pool to support custom workflows. If your user pool configuration does not include triggers, the ClientMetadata parameter serves no purpose.

Amazon Cognito does not validate the ClientMetadata value.

Amazon Cognito does not encrypt the the ClientMetadata value, so don't use it to provide sensitive information.

$sel:accessToken:GetUserAttributeVerificationCode', getUserAttributeVerificationCode_accessToken - The access token returned by the server response to get the user attribute verification code.

$sel:attributeName:GetUserAttributeVerificationCode', getUserAttributeVerificationCode_attributeName - The attribute name returned by the server response to get the user attribute verification code.

Request Lenses

getUserAttributeVerificationCode_clientMetadata :: Lens' GetUserAttributeVerificationCode (Maybe (HashMap Text Text))

A map of custom key-value pairs that you can provide as input for any custom workflows that this action triggers.

You create custom workflows by assigning Lambda functions to user pool triggers. When you use the GetUserAttributeVerificationCode API action, Amazon Cognito invokes the function that is assigned to the /custom message/ trigger. When Amazon Cognito invokes this function, it passes a JSON payload, which the function receives as input. This payload contains a clientMetadata attribute, which provides the data that you assigned to the ClientMetadata parameter in your GetUserAttributeVerificationCode request. In your function code in Lambda, you can process the clientMetadata value to enhance your workflow for your specific needs.

For more information, see Customizing User Pool Workflows with Lambda Triggers in the Amazon Cognito Developer Guide.

Take the following limitations into consideration when you use the ClientMetadata parameter:

Amazon Cognito does not store the ClientMetadata value. This data is available only to Lambda triggers that are assigned to a user pool to support custom workflows. If your user pool configuration does not include triggers, the ClientMetadata parameter serves no purpose.

Amazon Cognito does not validate the ClientMetadata value.

Amazon Cognito does not encrypt the the ClientMetadata value, so don't use it to provide sensitive information.

getUserAttributeVerificationCode_accessToken :: Lens' GetUserAttributeVerificationCode Text

The access token returned by the server response to get the user attribute verification code.

getUserAttributeVerificationCode_attributeName :: Lens' GetUserAttributeVerificationCode Text

The attribute name returned by the server response to get the user attribute verification code.

Destructuring the Response

data GetUserAttributeVerificationCodeResponse

The verification code response returned by the server response to get the user attribute verification code.

See: newGetUserAttributeVerificationCodeResponse smart constructor.

Constructors

GetUserAttributeVerificationCodeResponse'

Fields

codeDeliveryDetails :: Maybe CodeDeliveryDetailsType

The code delivery details returned by the server in response to the request to get the user attribute verification code.

httpStatus :: Int

The response's http status code.

newGetUserAttributeVerificationCodeResponse

Arguments

:: Int
$sel:httpStatus:GetUserAttributeVerificationCodeResponse'

-> GetUserAttributeVerificationCodeResponse

Create a value of GetUserAttributeVerificationCodeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:codeDeliveryDetails:GetUserAttributeVerificationCodeResponse', getUserAttributeVerificationCodeResponse_codeDeliveryDetails - The code delivery details returned by the server in response to the request to get the user attribute verification code.

$sel:httpStatus:GetUserAttributeVerificationCodeResponse', getUserAttributeVerificationCodeResponse_httpStatus - The response's http status code.

Response Lenses

getUserAttributeVerificationCodeResponse_codeDeliveryDetails :: Lens' GetUserAttributeVerificationCodeResponse (Maybe CodeDeliveryDetailsType)

The code delivery details returned by the server in response to the request to get the user attribute verification code.

getUserAttributeVerificationCodeResponse_httpStatus :: Lens' GetUserAttributeVerificationCodeResponse Int

The response's http status code.