Network.AWS.CognitoIdentityProvider.ResendConfirmationCode

Description

Resends the confirmation (for confirmation of registration) to a specific user in the user pool.

This action might generate an SMS text message. Starting June 1, 2021, U.S. telecom carriers require that you register an origination phone number before you can send SMS messages to U.S. phone numbers. If you use SMS text messages in Amazon Cognito, you must register a phone number with Amazon Pinpoint. Cognito will use the the registered number automatically. Otherwise, Cognito users that must receive SMS messages might be unable to sign up, activate their accounts, or sign in.

If you have never used SMS text messages with Amazon Cognito or any other Amazon Web Service, Amazon SNS might place your account in SMS sandbox. In <https:docs.aws.amazon.comsnslatestdgsns-sms-sandbox.html sandbox mode> , you’ll have limitations, such as sending messages to only verified phone numbers. After testing in the sandbox environment, you can move out of the SMS sandbox and into production. For more information, see SMS message settings for Cognito User Pools in the Amazon Cognito Developer Guide.

Creating a Request

data ResendConfirmationCode

Represents the request to resend the confirmation code.

See: newResendConfirmationCode smart constructor.

Constructors

ResendConfirmationCode'

Fields

clientMetadata :: Maybe (HashMap Text Text)

A map of custom key-value pairs that you can provide as input for any custom workflows that this action triggers.

You create custom workflows by assigning Lambda functions to user pool triggers. When you use the ResendConfirmationCode API action, Amazon Cognito invokes the function that is assigned to the custom message trigger. When Amazon Cognito invokes this function, it passes a JSON payload, which the function receives as input. This payload contains a clientMetadata attribute, which provides the data that you assigned to the ClientMetadata parameter in your ResendConfirmationCode request. In your function code in Lambda, you can process the clientMetadata value to enhance your workflow for your specific needs.

For more information, see Customizing User Pool Workflows with Lambda Triggers in the Amazon Cognito Developer Guide.

Take the following limitations into consideration when you use the ClientMetadata parameter:

Amazon Cognito does not store the ClientMetadata value. This data is available only to Lambda triggers that are assigned to a user pool to support custom workflows. If your user pool configuration does not include triggers, the ClientMetadata parameter serves no purpose.

Amazon Cognito does not validate the ClientMetadata value.

Amazon Cognito does not encrypt the the ClientMetadata value, so don't use it to provide sensitive information.

userContextData :: Maybe UserContextDataType

Contextual data such as the user's device fingerprint, IP address, or location used for evaluating the risk of an unexpected event by Amazon Cognito advanced security.

secretHash :: Maybe (Sensitive Text)

A keyed-hash message authentication code (HMAC) calculated using the secret key of a user pool client and username plus the client ID in the message.

analyticsMetadata :: Maybe AnalyticsMetadataType

The Amazon Pinpoint analytics metadata for collecting metrics for ResendConfirmationCode calls.

clientId :: Sensitive Text

The ID of the client associated with the user pool.

username :: Sensitive Text

The user name of the user to whom you wish to resend a confirmation code.

newResendConfirmationCode

Arguments

:: Text
$sel:clientId:ResendConfirmationCode'

-> Text
$sel:username:ResendConfirmationCode'

-> ResendConfirmationCode

Create a value of ResendConfirmationCode with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientMetadata:ResendConfirmationCode', resendConfirmationCode_clientMetadata - A map of custom key-value pairs that you can provide as input for any custom workflows that this action triggers.

You create custom workflows by assigning Lambda functions to user pool triggers. When you use the ResendConfirmationCode API action, Amazon Cognito invokes the function that is assigned to the custom message trigger. When Amazon Cognito invokes this function, it passes a JSON payload, which the function receives as input. This payload contains a clientMetadata attribute, which provides the data that you assigned to the ClientMetadata parameter in your ResendConfirmationCode request. In your function code in Lambda, you can process the clientMetadata value to enhance your workflow for your specific needs.

For more information, see Customizing User Pool Workflows with Lambda Triggers in the Amazon Cognito Developer Guide.

Take the following limitations into consideration when you use the ClientMetadata parameter:

Amazon Cognito does not store the ClientMetadata value. This data is available only to Lambda triggers that are assigned to a user pool to support custom workflows. If your user pool configuration does not include triggers, the ClientMetadata parameter serves no purpose.

Amazon Cognito does not validate the ClientMetadata value.

Amazon Cognito does not encrypt the the ClientMetadata value, so don't use it to provide sensitive information.

$sel:userContextData:ResendConfirmationCode', resendConfirmationCode_userContextData - Contextual data such as the user's device fingerprint, IP address, or location used for evaluating the risk of an unexpected event by Amazon Cognito advanced security.

$sel:secretHash:ResendConfirmationCode', resendConfirmationCode_secretHash - A keyed-hash message authentication code (HMAC) calculated using the secret key of a user pool client and username plus the client ID in the message.

$sel:analyticsMetadata:ResendConfirmationCode', resendConfirmationCode_analyticsMetadata - The Amazon Pinpoint analytics metadata for collecting metrics for ResendConfirmationCode calls.

$sel:clientId:ResendConfirmationCode', resendConfirmationCode_clientId - The ID of the client associated with the user pool.

$sel:username:ResendConfirmationCode', resendConfirmationCode_username - The user name of the user to whom you wish to resend a confirmation code.

Request Lenses

resendConfirmationCode_clientMetadata :: Lens' ResendConfirmationCode (Maybe (HashMap Text Text))

A map of custom key-value pairs that you can provide as input for any custom workflows that this action triggers.

You create custom workflows by assigning Lambda functions to user pool triggers. When you use the ResendConfirmationCode API action, Amazon Cognito invokes the function that is assigned to the custom message trigger. When Amazon Cognito invokes this function, it passes a JSON payload, which the function receives as input. This payload contains a clientMetadata attribute, which provides the data that you assigned to the ClientMetadata parameter in your ResendConfirmationCode request. In your function code in Lambda, you can process the clientMetadata value to enhance your workflow for your specific needs.

For more information, see Customizing User Pool Workflows with Lambda Triggers in the Amazon Cognito Developer Guide.

Take the following limitations into consideration when you use the ClientMetadata parameter:

Amazon Cognito does not store the ClientMetadata value. This data is available only to Lambda triggers that are assigned to a user pool to support custom workflows. If your user pool configuration does not include triggers, the ClientMetadata parameter serves no purpose.

Amazon Cognito does not validate the ClientMetadata value.

Amazon Cognito does not encrypt the the ClientMetadata value, so don't use it to provide sensitive information.

resendConfirmationCode_userContextData :: Lens' ResendConfirmationCode (Maybe UserContextDataType)

Contextual data such as the user's device fingerprint, IP address, or location used for evaluating the risk of an unexpected event by Amazon Cognito advanced security.

resendConfirmationCode_secretHash :: Lens' ResendConfirmationCode (Maybe Text)

A keyed-hash message authentication code (HMAC) calculated using the secret key of a user pool client and username plus the client ID in the message.

resendConfirmationCode_analyticsMetadata :: Lens' ResendConfirmationCode (Maybe AnalyticsMetadataType)

The Amazon Pinpoint analytics metadata for collecting metrics for ResendConfirmationCode calls.

resendConfirmationCode_clientId :: Lens' ResendConfirmationCode Text

The ID of the client associated with the user pool.

resendConfirmationCode_username :: Lens' ResendConfirmationCode Text

The user name of the user to whom you wish to resend a confirmation code.

Destructuring the Response

data ResendConfirmationCodeResponse

The response from the server when the Amazon Cognito Your User Pools service makes the request to resend a confirmation code.

See: newResendConfirmationCodeResponse smart constructor.

Constructors

ResendConfirmationCodeResponse'

Fields

codeDeliveryDetails :: Maybe CodeDeliveryDetailsType

The code delivery details returned by the server in response to the request to resend the confirmation code.

httpStatus :: Int

The response's http status code.

newResendConfirmationCodeResponse

Arguments

:: Int
$sel:httpStatus:ResendConfirmationCodeResponse'

-> ResendConfirmationCodeResponse

Create a value of ResendConfirmationCodeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:codeDeliveryDetails:ResendConfirmationCodeResponse', resendConfirmationCodeResponse_codeDeliveryDetails - The code delivery details returned by the server in response to the request to resend the confirmation code.

$sel:httpStatus:ResendConfirmationCodeResponse', resendConfirmationCodeResponse_httpStatus - The response's http status code.

Response Lenses

resendConfirmationCodeResponse_codeDeliveryDetails :: Lens' ResendConfirmationCodeResponse (Maybe CodeDeliveryDetailsType)

The code delivery details returned by the server in response to the request to resend the confirmation code.

resendConfirmationCodeResponse_httpStatus :: Lens' ResendConfirmationCodeResponse Int

The response's http status code.