Network.AWS.CognitoIdentityProvider.SetRiskConfiguration

Description

Configures actions on detected risks. To delete the risk configuration for UserPoolId or ClientId, pass null values for all four configuration types.

To enable Amazon Cognito advanced security features, update the user pool to include the UserPoolAddOns keyAdvancedSecurityMode.

Creating a Request

data SetRiskConfiguration

See: newSetRiskConfiguration smart constructor.

Constructors

SetRiskConfiguration'

Fields

riskExceptionConfiguration :: Maybe RiskExceptionConfigurationType

The configuration to override the risk decision.

clientId :: Maybe (Sensitive Text)

The app client ID. If ClientId is null, then the risk configuration is mapped to userPoolId. When the client ID is null, the same risk configuration is applied to all the clients in the userPool.

Otherwise, ClientId is mapped to the client. When the client ID is not null, the user pool configuration is overridden and the risk configuration for the client is used instead.

accountTakeoverRiskConfiguration :: Maybe AccountTakeoverRiskConfigurationType

The account takeover risk configuration.

compromisedCredentialsRiskConfiguration :: Maybe CompromisedCredentialsRiskConfigurationType

The compromised credentials risk configuration.

userPoolId :: Text

The user pool ID.

newSetRiskConfiguration

Arguments

:: Text
$sel:userPoolId:SetRiskConfiguration'

-> SetRiskConfiguration

Create a value of SetRiskConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:riskExceptionConfiguration:SetRiskConfiguration', setRiskConfiguration_riskExceptionConfiguration - The configuration to override the risk decision.

$sel:clientId:SetRiskConfiguration', setRiskConfiguration_clientId - The app client ID. If ClientId is null, then the risk configuration is mapped to userPoolId. When the client ID is null, the same risk configuration is applied to all the clients in the userPool.

Otherwise, ClientId is mapped to the client. When the client ID is not null, the user pool configuration is overridden and the risk configuration for the client is used instead.

$sel:accountTakeoverRiskConfiguration:SetRiskConfiguration', setRiskConfiguration_accountTakeoverRiskConfiguration - The account takeover risk configuration.

$sel:compromisedCredentialsRiskConfiguration:SetRiskConfiguration', setRiskConfiguration_compromisedCredentialsRiskConfiguration - The compromised credentials risk configuration.

$sel:userPoolId:SetRiskConfiguration', setRiskConfiguration_userPoolId - The user pool ID.

Request Lenses

setRiskConfiguration_riskExceptionConfiguration :: Lens' SetRiskConfiguration (Maybe RiskExceptionConfigurationType)

The configuration to override the risk decision.

setRiskConfiguration_clientId :: Lens' SetRiskConfiguration (Maybe Text)

The app client ID. If ClientId is null, then the risk configuration is mapped to userPoolId. When the client ID is null, the same risk configuration is applied to all the clients in the userPool.

Otherwise, ClientId is mapped to the client. When the client ID is not null, the user pool configuration is overridden and the risk configuration for the client is used instead.

setRiskConfiguration_accountTakeoverRiskConfiguration :: Lens' SetRiskConfiguration (Maybe AccountTakeoverRiskConfigurationType)

The account takeover risk configuration.

setRiskConfiguration_compromisedCredentialsRiskConfiguration :: Lens' SetRiskConfiguration (Maybe CompromisedCredentialsRiskConfigurationType)

The compromised credentials risk configuration.

setRiskConfiguration_userPoolId :: Lens' SetRiskConfiguration Text

The user pool ID.

Destructuring the Response

data SetRiskConfigurationResponse

See: newSetRiskConfigurationResponse smart constructor.

Constructors

SetRiskConfigurationResponse'

Fields

httpStatus :: Int

The response's http status code.

riskConfiguration :: RiskConfigurationType

The risk configuration.

newSetRiskConfigurationResponse

Arguments

:: Int
$sel:httpStatus:SetRiskConfigurationResponse'

-> RiskConfigurationType
$sel:riskConfiguration:SetRiskConfigurationResponse'

-> SetRiskConfigurationResponse

Create a value of SetRiskConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:SetRiskConfigurationResponse', setRiskConfigurationResponse_httpStatus - The response's http status code.

$sel:riskConfiguration:SetRiskConfigurationResponse', setRiskConfigurationResponse_riskConfiguration - The risk configuration.

Response Lenses

setRiskConfigurationResponse_httpStatus :: Lens' SetRiskConfigurationResponse Int

The response's http status code.

setRiskConfigurationResponse_riskConfiguration :: Lens' SetRiskConfigurationResponse RiskConfigurationType

The risk configuration.