Network.AWS.CognitoIdentityProvider.SetUserMFAPreference

Description

Set the user's multi-factor authentication (MFA) method preference, including which MFA factors are enabled and if any are preferred. Only one factor can be set as preferred. The preferred MFA factor will be used to authenticate a user if multiple factors are enabled. If multiple options are enabled and no preference is set, a challenge to choose an MFA option will be returned during sign in. If an MFA type is enabled for a user, the user will be prompted for MFA during all sign in attempts, unless device tracking is turned on and the device has been trusted. If you would like MFA to be applied selectively based on the assessed risk level of sign in attempts, disable MFA for users and turn on Adaptive Authentication for the user pool.

Creating a Request

data SetUserMFAPreference

See: newSetUserMFAPreference smart constructor.

Constructors

SetUserMFAPreference'

Fields

softwareTokenMfaSettings :: Maybe SoftwareTokenMfaSettingsType

The time-based one-time password software token MFA settings.

sMSMfaSettings :: Maybe SMSMfaSettingsType

The SMS text message multi-factor authentication (MFA) settings.

accessToken :: Sensitive Text

The access token for the user.

newSetUserMFAPreference

Arguments

:: Text
$sel:accessToken:SetUserMFAPreference'

-> SetUserMFAPreference

Create a value of SetUserMFAPreference with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:softwareTokenMfaSettings:SetUserMFAPreference', setUserMFAPreference_softwareTokenMfaSettings - The time-based one-time password software token MFA settings.

$sel:sMSMfaSettings:SetUserMFAPreference', setUserMFAPreference_sMSMfaSettings - The SMS text message multi-factor authentication (MFA) settings.

$sel:accessToken:SetUserMFAPreference', setUserMFAPreference_accessToken - The access token for the user.

Request Lenses

setUserMFAPreference_softwareTokenMfaSettings :: Lens' SetUserMFAPreference (Maybe SoftwareTokenMfaSettingsType)

The time-based one-time password software token MFA settings.

setUserMFAPreference_sMSMfaSettings :: Lens' SetUserMFAPreference (Maybe SMSMfaSettingsType)

The SMS text message multi-factor authentication (MFA) settings.

setUserMFAPreference_accessToken :: Lens' SetUserMFAPreference Text

The access token for the user.

Destructuring the Response

data SetUserMFAPreferenceResponse

See: newSetUserMFAPreferenceResponse smart constructor.

Constructors

SetUserMFAPreferenceResponse'

Fields

httpStatus :: Int

The response's http status code.

newSetUserMFAPreferenceResponse

Arguments

:: Int
$sel:httpStatus:SetUserMFAPreferenceResponse'

-> SetUserMFAPreferenceResponse

Create a value of SetUserMFAPreferenceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:SetUserMFAPreferenceResponse', setUserMFAPreferenceResponse_httpStatus - The response's http status code.

Response Lenses

setUserMFAPreferenceResponse_httpStatus :: Lens' SetUserMFAPreferenceResponse Int

The response's http status code.