Network.AWS.CognitoIdentityProvider.SetUserPoolMfaConfig

Description

Set the user pool multi-factor authentication (MFA) configuration.

This action might generate an SMS text message. Starting June 1, 2021, U.S. telecom carriers require that you register an origination phone number before you can send SMS messages to U.S. phone numbers. If you use SMS text messages in Amazon Cognito, you must register a phone number with Amazon Pinpoint. Cognito will use the the registered number automatically. Otherwise, Cognito users that must receive SMS messages might be unable to sign up, activate their accounts, or sign in.

If you have never used SMS text messages with Amazon Cognito or any other Amazon Web Service, Amazon SNS might place your account in SMS sandbox. In <https:docs.aws.amazon.comsnslatestdgsns-sms-sandbox.html sandbox mode> , you’ll have limitations, such as sending messages to only verified phone numbers. After testing in the sandbox environment, you can move out of the SMS sandbox and into production. For more information, see SMS message settings for Cognito User Pools in the Amazon Cognito Developer Guide.

Creating a Request

data SetUserPoolMfaConfig

See: newSetUserPoolMfaConfig smart constructor.

Constructors

SetUserPoolMfaConfig'

Fields

softwareTokenMfaConfiguration :: Maybe SoftwareTokenMfaConfigType

The software token MFA configuration.

smsMfaConfiguration :: Maybe SmsMfaConfigType

The SMS text message MFA configuration.

mfaConfiguration :: Maybe UserPoolMfaType

The MFA configuration. Users who don't have an MFA factor set up won't be able to sign-in if you set the MfaConfiguration value to ‘ON’. See Adding Multi-Factor Authentication (MFA) to a User Pool to learn more. Valid values include:

OFF MFA will not be used for any users.

ON MFA is required for all users to sign in.

OPTIONAL MFA will be required only for individual users who have an MFA factor enabled.

userPoolId :: Text

The user pool ID.

newSetUserPoolMfaConfig

Arguments

:: Text
$sel:userPoolId:SetUserPoolMfaConfig'

-> SetUserPoolMfaConfig

Create a value of SetUserPoolMfaConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:softwareTokenMfaConfiguration:SetUserPoolMfaConfig', setUserPoolMfaConfig_softwareTokenMfaConfiguration - The software token MFA configuration.

$sel:smsMfaConfiguration:SetUserPoolMfaConfig', setUserPoolMfaConfig_smsMfaConfiguration - The SMS text message MFA configuration.

$sel:mfaConfiguration:SetUserPoolMfaConfig', setUserPoolMfaConfig_mfaConfiguration - The MFA configuration. Users who don't have an MFA factor set up won't be able to sign-in if you set the MfaConfiguration value to ‘ON’. See Adding Multi-Factor Authentication (MFA) to a User Pool to learn more. Valid values include:

OFF MFA will not be used for any users.

ON MFA is required for all users to sign in.

OPTIONAL MFA will be required only for individual users who have an MFA factor enabled.

$sel:userPoolId:SetUserPoolMfaConfig', setUserPoolMfaConfig_userPoolId - The user pool ID.

Request Lenses

setUserPoolMfaConfig_softwareTokenMfaConfiguration :: Lens' SetUserPoolMfaConfig (Maybe SoftwareTokenMfaConfigType)

The software token MFA configuration.

setUserPoolMfaConfig_smsMfaConfiguration :: Lens' SetUserPoolMfaConfig (Maybe SmsMfaConfigType)

The SMS text message MFA configuration.

setUserPoolMfaConfig_mfaConfiguration :: Lens' SetUserPoolMfaConfig (Maybe UserPoolMfaType)

The MFA configuration. Users who don't have an MFA factor set up won't be able to sign-in if you set the MfaConfiguration value to ‘ON’. See Adding Multi-Factor Authentication (MFA) to a User Pool to learn more. Valid values include:

OFF MFA will not be used for any users.

ON MFA is required for all users to sign in.

OPTIONAL MFA will be required only for individual users who have an MFA factor enabled.

setUserPoolMfaConfig_userPoolId :: Lens' SetUserPoolMfaConfig Text

The user pool ID.

Destructuring the Response

data SetUserPoolMfaConfigResponse

See: newSetUserPoolMfaConfigResponse smart constructor.

Constructors

SetUserPoolMfaConfigResponse'

Fields

softwareTokenMfaConfiguration :: Maybe SoftwareTokenMfaConfigType

The software token MFA configuration.

smsMfaConfiguration :: Maybe SmsMfaConfigType

The SMS text message MFA configuration.

mfaConfiguration :: Maybe UserPoolMfaType

The MFA configuration. Valid values include:

OFF MFA will not be used for any users.

ON MFA is required for all users to sign in.

OPTIONAL MFA will be required only for individual users who have an MFA factor enabled.

httpStatus :: Int

The response's http status code.

newSetUserPoolMfaConfigResponse

Arguments

:: Int
$sel:httpStatus:SetUserPoolMfaConfigResponse'

-> SetUserPoolMfaConfigResponse

Create a value of SetUserPoolMfaConfigResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:softwareTokenMfaConfiguration:SetUserPoolMfaConfig', setUserPoolMfaConfigResponse_softwareTokenMfaConfiguration - The software token MFA configuration.

$sel:smsMfaConfiguration:SetUserPoolMfaConfig', setUserPoolMfaConfigResponse_smsMfaConfiguration - The SMS text message MFA configuration.

$sel:mfaConfiguration:SetUserPoolMfaConfig', setUserPoolMfaConfigResponse_mfaConfiguration - The MFA configuration. Valid values include:

OFF MFA will not be used for any users.

ON MFA is required for all users to sign in.

OPTIONAL MFA will be required only for individual users who have an MFA factor enabled.

$sel:httpStatus:SetUserPoolMfaConfigResponse', setUserPoolMfaConfigResponse_httpStatus - The response's http status code.

Response Lenses

setUserPoolMfaConfigResponse_softwareTokenMfaConfiguration :: Lens' SetUserPoolMfaConfigResponse (Maybe SoftwareTokenMfaConfigType)

The software token MFA configuration.

setUserPoolMfaConfigResponse_smsMfaConfiguration :: Lens' SetUserPoolMfaConfigResponse (Maybe SmsMfaConfigType)

The SMS text message MFA configuration.

setUserPoolMfaConfigResponse_mfaConfiguration :: Lens' SetUserPoolMfaConfigResponse (Maybe UserPoolMfaType)

The MFA configuration. Valid values include:

OFF MFA will not be used for any users.

ON MFA is required for all users to sign in.

OPTIONAL MFA will be required only for individual users who have an MFA factor enabled.

setUserPoolMfaConfigResponse_httpStatus :: Lens' SetUserPoolMfaConfigResponse Int

The response's http status code.