Network.AWS.CognitoIdentityProvider.SignUp

Description

Registers the user in the specified user pool and creates a user name, password, and user attributes.

This action might generate an SMS text message. Starting June 1, 2021, U.S. telecom carriers require that you register an origination phone number before you can send SMS messages to U.S. phone numbers. If you use SMS text messages in Amazon Cognito, you must register a phone number with Amazon Pinpoint. Cognito will use the the registered number automatically. Otherwise, Cognito users that must receive SMS messages might be unable to sign up, activate their accounts, or sign in.

If you have never used SMS text messages with Amazon Cognito or any other Amazon Web Service, Amazon SNS might place your account in SMS sandbox. In <https:docs.aws.amazon.comsnslatestdgsns-sms-sandbox.html sandbox mode> , you’ll have limitations, such as sending messages to only verified phone numbers. After testing in the sandbox environment, you can move out of the SMS sandbox and into production. For more information, see SMS message settings for Cognito User Pools in the Amazon Cognito Developer Guide.

Creating a Request

data SignUp

Represents the request to register a user.

See: newSignUp smart constructor.

Constructors

SignUp'

Fields

clientMetadata :: Maybe (HashMap Text Text)

A map of custom key-value pairs that you can provide as input for any custom workflows that this action triggers.

You create custom workflows by assigning Lambda functions to user pool triggers. When you use the SignUp API action, Amazon Cognito invokes any functions that are assigned to the following triggers: pre sign-up, custom message, and post confirmation. When Amazon Cognito invokes any of these functions, it passes a JSON payload, which the function receives as input. This payload contains a clientMetadata attribute, which provides the data that you assigned to the ClientMetadata parameter in your SignUp request. In your function code in Lambda, you can process the clientMetadata value to enhance your workflow for your specific needs.

For more information, see Customizing User Pool Workflows with Lambda Triggers in the Amazon Cognito Developer Guide.

Take the following limitations into consideration when you use the ClientMetadata parameter:

Amazon Cognito does not store the ClientMetadata value. This data is available only to Lambda triggers that are assigned to a user pool to support custom workflows. If your user pool configuration does not include triggers, the ClientMetadata parameter serves no purpose.

Amazon Cognito does not validate the ClientMetadata value.

Amazon Cognito does not encrypt the the ClientMetadata value, so don't use it to provide sensitive information.

userContextData :: Maybe UserContextDataType

Contextual data such as the user's device fingerprint, IP address, or location used for evaluating the risk of an unexpected event by Amazon Cognito advanced security.

secretHash :: Maybe (Sensitive Text)

A keyed-hash message authentication code (HMAC) calculated using the secret key of a user pool client and username plus the client ID in the message.

userAttributes :: Maybe [AttributeType]

An array of name-value pairs representing user attributes.

For custom attributes, you must prepend the custom: prefix to the attribute name.

analyticsMetadata :: Maybe AnalyticsMetadataType

The Amazon Pinpoint analytics metadata for collecting metrics for SignUp calls.

validationData :: Maybe [AttributeType]

The validation data in the request to register a user.

clientId :: Sensitive Text

The ID of the client associated with the user pool.

username :: Sensitive Text

The user name of the user you wish to register.

password :: Sensitive Text

The password of the user you wish to register.

newSignUp

Arguments

:: Text
$sel:clientId:SignUp'

-> Text
$sel:username:SignUp'

-> Text
$sel:password:SignUp'

-> SignUp

Create a value of SignUp with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientMetadata:SignUp', signUp_clientMetadata - A map of custom key-value pairs that you can provide as input for any custom workflows that this action triggers.

You create custom workflows by assigning Lambda functions to user pool triggers. When you use the SignUp API action, Amazon Cognito invokes any functions that are assigned to the following triggers: pre sign-up, custom message, and post confirmation. When Amazon Cognito invokes any of these functions, it passes a JSON payload, which the function receives as input. This payload contains a clientMetadata attribute, which provides the data that you assigned to the ClientMetadata parameter in your SignUp request. In your function code in Lambda, you can process the clientMetadata value to enhance your workflow for your specific needs.

For more information, see Customizing User Pool Workflows with Lambda Triggers in the Amazon Cognito Developer Guide.

Take the following limitations into consideration when you use the ClientMetadata parameter:

Amazon Cognito does not store the ClientMetadata value. This data is available only to Lambda triggers that are assigned to a user pool to support custom workflows. If your user pool configuration does not include triggers, the ClientMetadata parameter serves no purpose.

Amazon Cognito does not validate the ClientMetadata value.

Amazon Cognito does not encrypt the the ClientMetadata value, so don't use it to provide sensitive information.

$sel:userContextData:SignUp', signUp_userContextData - Contextual data such as the user's device fingerprint, IP address, or location used for evaluating the risk of an unexpected event by Amazon Cognito advanced security.

$sel:secretHash:SignUp', signUp_secretHash - A keyed-hash message authentication code (HMAC) calculated using the secret key of a user pool client and username plus the client ID in the message.

$sel:userAttributes:SignUp', signUp_userAttributes - An array of name-value pairs representing user attributes.

For custom attributes, you must prepend the custom: prefix to the attribute name.

$sel:analyticsMetadata:SignUp', signUp_analyticsMetadata - The Amazon Pinpoint analytics metadata for collecting metrics for SignUp calls.

$sel:validationData:SignUp', signUp_validationData - The validation data in the request to register a user.

$sel:clientId:SignUp', signUp_clientId - The ID of the client associated with the user pool.

$sel:username:SignUp', signUp_username - The user name of the user you wish to register.

$sel:password:SignUp', signUp_password - The password of the user you wish to register.

Request Lenses

signUp_clientMetadata :: Lens' SignUp (Maybe (HashMap Text Text))

A map of custom key-value pairs that you can provide as input for any custom workflows that this action triggers.

You create custom workflows by assigning Lambda functions to user pool triggers. When you use the SignUp API action, Amazon Cognito invokes any functions that are assigned to the following triggers: pre sign-up, custom message, and post confirmation. When Amazon Cognito invokes any of these functions, it passes a JSON payload, which the function receives as input. This payload contains a clientMetadata attribute, which provides the data that you assigned to the ClientMetadata parameter in your SignUp request. In your function code in Lambda, you can process the clientMetadata value to enhance your workflow for your specific needs.

For more information, see Customizing User Pool Workflows with Lambda Triggers in the Amazon Cognito Developer Guide.

Take the following limitations into consideration when you use the ClientMetadata parameter:

Amazon Cognito does not store the ClientMetadata value. This data is available only to Lambda triggers that are assigned to a user pool to support custom workflows. If your user pool configuration does not include triggers, the ClientMetadata parameter serves no purpose.

Amazon Cognito does not validate the ClientMetadata value.

Amazon Cognito does not encrypt the the ClientMetadata value, so don't use it to provide sensitive information.

signUp_userContextData :: Lens' SignUp (Maybe UserContextDataType)

Contextual data such as the user's device fingerprint, IP address, or location used for evaluating the risk of an unexpected event by Amazon Cognito advanced security.

signUp_secretHash :: Lens' SignUp (Maybe Text)

A keyed-hash message authentication code (HMAC) calculated using the secret key of a user pool client and username plus the client ID in the message.

signUp_userAttributes :: Lens' SignUp (Maybe [AttributeType])

An array of name-value pairs representing user attributes.

For custom attributes, you must prepend the custom: prefix to the attribute name.

signUp_analyticsMetadata :: Lens' SignUp (Maybe AnalyticsMetadataType)

The Amazon Pinpoint analytics metadata for collecting metrics for SignUp calls.

signUp_validationData :: Lens' SignUp (Maybe [AttributeType])

The validation data in the request to register a user.

signUp_clientId :: Lens' SignUp Text

The ID of the client associated with the user pool.

signUp_username :: Lens' SignUp Text

The user name of the user you wish to register.

signUp_password :: Lens' SignUp Text

The password of the user you wish to register.

Destructuring the Response

data SignUpResponse

The response from the server for a registration request.

See: newSignUpResponse smart constructor.

Constructors

SignUpResponse'

Fields

codeDeliveryDetails :: Maybe CodeDeliveryDetailsType

The code delivery details returned by the server response to the user registration request.

httpStatus :: Int

The response's http status code.

userConfirmed :: Bool

A response from the server indicating that a user registration has been confirmed.

userSub :: Text

The UUID of the authenticated user. This is not the same as username.

newSignUpResponse

Arguments

:: Int
$sel:httpStatus:SignUpResponse'

-> Bool
$sel:userConfirmed:SignUpResponse'

-> Text
$sel:userSub:SignUpResponse'

-> SignUpResponse

Create a value of SignUpResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:codeDeliveryDetails:SignUpResponse', signUpResponse_codeDeliveryDetails - The code delivery details returned by the server response to the user registration request.

$sel:httpStatus:SignUpResponse', signUpResponse_httpStatus - The response's http status code.

$sel:userConfirmed:SignUpResponse', signUpResponse_userConfirmed - A response from the server indicating that a user registration has been confirmed.

$sel:userSub:SignUpResponse', signUpResponse_userSub - The UUID of the authenticated user. This is not the same as username.

Response Lenses

signUpResponse_codeDeliveryDetails :: Lens' SignUpResponse (Maybe CodeDeliveryDetailsType)

The code delivery details returned by the server response to the user registration request.

signUpResponse_httpStatus :: Lens' SignUpResponse Int

The response's http status code.

signUpResponse_userConfirmed :: Lens' SignUpResponse Bool

A response from the server indicating that a user registration has been confirmed.

signUpResponse_userSub :: Lens' SignUpResponse Text

The UUID of the authenticated user. This is not the same as username.