Network.AWS.CognitoIdentityProvider.StopUserImportJob

Description

Stops the user import job.

Creating a Request

data StopUserImportJob

Represents the request to stop the user import job.

See: newStopUserImportJob smart constructor.

Constructors

StopUserImportJob'

Fields

userPoolId :: Text

The user pool ID for the user pool that the users are being imported into.

jobId :: Text

The job ID for the user import job.

newStopUserImportJob

Arguments

:: Text
$sel:userPoolId:StopUserImportJob'

-> Text
$sel:jobId:StopUserImportJob'

-> StopUserImportJob

Create a value of StopUserImportJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userPoolId:StopUserImportJob', stopUserImportJob_userPoolId - The user pool ID for the user pool that the users are being imported into.

$sel:jobId:StopUserImportJob', stopUserImportJob_jobId - The job ID for the user import job.

Request Lenses

stopUserImportJob_userPoolId :: Lens' StopUserImportJob Text

The user pool ID for the user pool that the users are being imported into.

stopUserImportJob_jobId :: Lens' StopUserImportJob Text

The job ID for the user import job.

Destructuring the Response

data StopUserImportJobResponse

Represents the response from the server to the request to stop the user import job.

See: newStopUserImportJobResponse smart constructor.

Constructors

StopUserImportJobResponse'

Fields

userImportJob :: Maybe UserImportJobType

The job object that represents the user import job.

httpStatus :: Int

The response's http status code.

newStopUserImportJobResponse

Arguments

:: Int
$sel:httpStatus:StopUserImportJobResponse'

-> StopUserImportJobResponse

Create a value of StopUserImportJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userImportJob:StopUserImportJobResponse', stopUserImportJobResponse_userImportJob - The job object that represents the user import job.

$sel:httpStatus:StopUserImportJobResponse', stopUserImportJobResponse_httpStatus - The response's http status code.

Response Lenses

stopUserImportJobResponse_userImportJob :: Lens' StopUserImportJobResponse (Maybe UserImportJobType)

The job object that represents the user import job.

stopUserImportJobResponse_httpStatus :: Lens' StopUserImportJobResponse Int

The response's http status code.