Network.AWS.CognitoIdentityProvider.Types.AnalyticsConfigurationType

Description

Documentation

data AnalyticsConfigurationType

The Amazon Pinpoint analytics configuration for collecting metrics for a user pool.

In regions where Pinpoint is not available, Cognito User Pools only supports sending events to Amazon Pinpoint projects in us-east-1. In regions where Pinpoint is available, Cognito User Pools will support sending events to Amazon Pinpoint projects within that same region.

See: newAnalyticsConfigurationType smart constructor.

Constructors

AnalyticsConfigurationType'

Fields

applicationArn :: Maybe Text

The Amazon Resource Name (ARN) of an Amazon Pinpoint project. You can use the Amazon Pinpoint project for Pinpoint integration with the chosen User Pool Client. Amazon Cognito publishes events to the pinpoint project declared by the app ARN.

userDataShared :: Maybe Bool

If UserDataShared is true, Amazon Cognito will include user data in the events it publishes to Amazon Pinpoint analytics.

applicationId :: Maybe Text

The application ID for an Amazon Pinpoint application.

externalId :: Maybe Text

The external ID.

roleArn :: Maybe Text

The ARN of an IAM role that authorizes Amazon Cognito to publish events to Amazon Pinpoint analytics.

newAnalyticsConfigurationType :: AnalyticsConfigurationType

Create a value of AnalyticsConfigurationType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationArn:AnalyticsConfigurationType', analyticsConfigurationType_applicationArn - The Amazon Resource Name (ARN) of an Amazon Pinpoint project. You can use the Amazon Pinpoint project for Pinpoint integration with the chosen User Pool Client. Amazon Cognito publishes events to the pinpoint project declared by the app ARN.

$sel:userDataShared:AnalyticsConfigurationType', analyticsConfigurationType_userDataShared - If UserDataShared is true, Amazon Cognito will include user data in the events it publishes to Amazon Pinpoint analytics.

$sel:applicationId:AnalyticsConfigurationType', analyticsConfigurationType_applicationId - The application ID for an Amazon Pinpoint application.

$sel:externalId:AnalyticsConfigurationType', analyticsConfigurationType_externalId - The external ID.

$sel:roleArn:AnalyticsConfigurationType', analyticsConfigurationType_roleArn - The ARN of an IAM role that authorizes Amazon Cognito to publish events to Amazon Pinpoint analytics.

analyticsConfigurationType_applicationArn :: Lens' AnalyticsConfigurationType (Maybe Text)

The Amazon Resource Name (ARN) of an Amazon Pinpoint project. You can use the Amazon Pinpoint project for Pinpoint integration with the chosen User Pool Client. Amazon Cognito publishes events to the pinpoint project declared by the app ARN.

analyticsConfigurationType_userDataShared :: Lens' AnalyticsConfigurationType (Maybe Bool)

If UserDataShared is true, Amazon Cognito will include user data in the events it publishes to Amazon Pinpoint analytics.

analyticsConfigurationType_applicationId :: Lens' AnalyticsConfigurationType (Maybe Text)

The application ID for an Amazon Pinpoint application.

analyticsConfigurationType_externalId :: Lens' AnalyticsConfigurationType (Maybe Text)

The external ID.

analyticsConfigurationType_roleArn :: Lens' AnalyticsConfigurationType (Maybe Text)

The ARN of an IAM role that authorizes Amazon Cognito to publish events to Amazon Pinpoint analytics.