Network.AWS.CognitoIdentityProvider.Types.AuthEventType

Description

Documentation

data AuthEventType

The authentication event type.

See: newAuthEventType smart constructor.

Constructors

AuthEventType'

Fields

eventType :: Maybe EventType

The event type.

eventId :: Maybe Text

The event ID.

challengeResponses :: Maybe [ChallengeResponseType]

The challenge responses.

eventContextData :: Maybe EventContextDataType

The user context data captured at the time of an event request. It provides additional information about the client from which event the request is received.

creationDate :: Maybe POSIX

The creation date

eventRisk :: Maybe EventRiskType

The event risk.

eventResponse :: Maybe EventResponseType

The event response.

eventFeedback :: Maybe EventFeedbackType

A flag specifying the user feedback captured at the time of an event request is good or bad.

newAuthEventType :: AuthEventType

Create a value of AuthEventType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventType:AuthEventType', authEventType_eventType - The event type.

$sel:eventId:AuthEventType', authEventType_eventId - The event ID.

$sel:challengeResponses:AuthEventType', authEventType_challengeResponses - The challenge responses.

$sel:eventContextData:AuthEventType', authEventType_eventContextData - The user context data captured at the time of an event request. It provides additional information about the client from which event the request is received.

$sel:creationDate:AuthEventType', authEventType_creationDate - The creation date

$sel:eventRisk:AuthEventType', authEventType_eventRisk - The event risk.

$sel:eventResponse:AuthEventType', authEventType_eventResponse - The event response.

$sel:eventFeedback:AuthEventType', authEventType_eventFeedback - A flag specifying the user feedback captured at the time of an event request is good or bad.

authEventType_eventType :: Lens' AuthEventType (Maybe EventType)

The event type.

authEventType_eventId :: Lens' AuthEventType (Maybe Text)

The event ID.

authEventType_challengeResponses :: Lens' AuthEventType (Maybe [ChallengeResponseType])

The challenge responses.

authEventType_eventContextData :: Lens' AuthEventType (Maybe EventContextDataType)

The user context data captured at the time of an event request. It provides additional information about the client from which event the request is received.

authEventType_creationDate :: Lens' AuthEventType (Maybe UTCTime)

The creation date

authEventType_eventRisk :: Lens' AuthEventType (Maybe EventRiskType)

The event risk.

authEventType_eventResponse :: Lens' AuthEventType (Maybe EventResponseType)

The event response.

authEventType_eventFeedback :: Lens' AuthEventType (Maybe EventFeedbackType)

A flag specifying the user feedback captured at the time of an event request is good or bad.