Network.AWS.CognitoIdentityProvider.Types.AuthenticationResultType

Description

Documentation

data AuthenticationResultType

The authentication result.

See: newAuthenticationResultType smart constructor.

Constructors

AuthenticationResultType'

Fields

expiresIn :: Maybe Int

The expiration period of the authentication result in seconds.

tokenType :: Maybe Text

The token type.

accessToken :: Maybe (Sensitive Text)

The access token.

newDeviceMetadata' :: Maybe NewDeviceMetadataType

The new device metadata from an authentication result.

refreshToken :: Maybe (Sensitive Text)

The refresh token.

idToken :: Maybe (Sensitive Text)

The ID token.

newAuthenticationResultType :: AuthenticationResultType

Create a value of AuthenticationResultType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expiresIn:AuthenticationResultType', authenticationResultType_expiresIn - The expiration period of the authentication result in seconds.

$sel:tokenType:AuthenticationResultType', authenticationResultType_tokenType - The token type.

$sel:accessToken:AuthenticationResultType', authenticationResultType_accessToken - The access token.

$sel:newDeviceMetadata':AuthenticationResultType', authenticationResultType_newDeviceMetadata - The new device metadata from an authentication result.

$sel:refreshToken:AuthenticationResultType', authenticationResultType_refreshToken - The refresh token.

$sel:idToken:AuthenticationResultType', authenticationResultType_idToken - The ID token.

authenticationResultType_expiresIn :: Lens' AuthenticationResultType (Maybe Int)

The expiration period of the authentication result in seconds.

authenticationResultType_tokenType :: Lens' AuthenticationResultType (Maybe Text)

The token type.

authenticationResultType_accessToken :: Lens' AuthenticationResultType (Maybe Text)

The access token.

authenticationResultType_newDeviceMetadata :: Lens' AuthenticationResultType (Maybe NewDeviceMetadataType)

The new device metadata from an authentication result.

authenticationResultType_refreshToken :: Lens' AuthenticationResultType (Maybe Text)

The refresh token.

authenticationResultType_idToken :: Lens' AuthenticationResultType (Maybe Text)

The ID token.