Network.AWS.CognitoIdentityProvider.Types.DomainDescriptionType

Description

Documentation

data DomainDescriptionType

A container for information about a domain.

See: newDomainDescriptionType smart constructor.

Constructors

DomainDescriptionType'

Fields

status :: Maybe DomainStatusType

The domain status.

aWSAccountId :: Maybe Text

The account ID for the user pool owner.

customDomainConfig :: Maybe CustomDomainConfigType

The configuration for a custom domain that hosts the sign-up and sign-in webpages for your application.

s3Bucket :: Maybe Text

The S3 bucket where the static files for this domain are stored.

userPoolId :: Maybe Text

The user pool ID.

domain :: Maybe Text

The domain string.

version :: Maybe Text

The app version.

cloudFrontDistribution :: Maybe Text

The ARN of the CloudFront distribution.

newDomainDescriptionType :: DomainDescriptionType

Create a value of DomainDescriptionType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DomainDescriptionType', domainDescriptionType_status - The domain status.

$sel:aWSAccountId:DomainDescriptionType', domainDescriptionType_aWSAccountId - The account ID for the user pool owner.

$sel:customDomainConfig:DomainDescriptionType', domainDescriptionType_customDomainConfig - The configuration for a custom domain that hosts the sign-up and sign-in webpages for your application.

$sel:s3Bucket:DomainDescriptionType', domainDescriptionType_s3Bucket - The S3 bucket where the static files for this domain are stored.

$sel:userPoolId:DomainDescriptionType', domainDescriptionType_userPoolId - The user pool ID.

$sel:domain:DomainDescriptionType', domainDescriptionType_domain - The domain string.

$sel:version:DomainDescriptionType', domainDescriptionType_version - The app version.

$sel:cloudFrontDistribution:DomainDescriptionType', domainDescriptionType_cloudFrontDistribution - The ARN of the CloudFront distribution.

domainDescriptionType_status :: Lens' DomainDescriptionType (Maybe DomainStatusType)

The domain status.

domainDescriptionType_aWSAccountId :: Lens' DomainDescriptionType (Maybe Text)

The account ID for the user pool owner.

domainDescriptionType_customDomainConfig :: Lens' DomainDescriptionType (Maybe CustomDomainConfigType)

The configuration for a custom domain that hosts the sign-up and sign-in webpages for your application.

domainDescriptionType_s3Bucket :: Lens' DomainDescriptionType (Maybe Text)

The S3 bucket where the static files for this domain are stored.

domainDescriptionType_userPoolId :: Lens' DomainDescriptionType (Maybe Text)

The user pool ID.

domainDescriptionType_domain :: Lens' DomainDescriptionType (Maybe Text)

The domain string.

domainDescriptionType_version :: Lens' DomainDescriptionType (Maybe Text)

The app version.

domainDescriptionType_cloudFrontDistribution :: Lens' DomainDescriptionType (Maybe Text)

The ARN of the CloudFront distribution.