Network.AWS.CognitoIdentityProvider.Types.GroupType

Description

Documentation

data GroupType

The group type.

See: newGroupType smart constructor.

Constructors

GroupType'

Fields

lastModifiedDate :: Maybe POSIX

The date the group was last modified.

roleArn :: Maybe Text

The role ARN for the group.

userPoolId :: Maybe Text

The user pool ID for the user pool.

groupName :: Maybe Text

The name of the group.

creationDate :: Maybe POSIX

The date the group was created.

description :: Maybe Text

A string containing the description of the group.

precedence :: Maybe Natural

A nonnegative integer value that specifies the precedence of this group relative to the other groups that a user can belong to in the user pool. If a user belongs to two or more groups, it is the group with the highest precedence whose role ARN will be used in the cognito:roles and cognito:preferred_role claims in the user's tokens. Groups with higher Precedence values take precedence over groups with lower Precedence values or with null Precedence values.

Two groups can have the same Precedence value. If this happens, neither group takes precedence over the other. If two groups with the same Precedence have the same role ARN, that role is used in the cognito:preferred_role claim in tokens for users in each group. If the two groups have different role ARNs, the cognito:preferred_role claim is not set in users' tokens.

The default Precedence value is null.

newGroupType :: GroupType

Create a value of GroupType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedDate:GroupType', groupType_lastModifiedDate - The date the group was last modified.

$sel:roleArn:GroupType', groupType_roleArn - The role ARN for the group.

$sel:userPoolId:GroupType', groupType_userPoolId - The user pool ID for the user pool.

$sel:groupName:GroupType', groupType_groupName - The name of the group.

$sel:creationDate:GroupType', groupType_creationDate - The date the group was created.

$sel:description:GroupType', groupType_description - A string containing the description of the group.

$sel:precedence:GroupType', groupType_precedence - A nonnegative integer value that specifies the precedence of this group relative to the other groups that a user can belong to in the user pool. If a user belongs to two or more groups, it is the group with the highest precedence whose role ARN will be used in the cognito:roles and cognito:preferred_role claims in the user's tokens. Groups with higher Precedence values take precedence over groups with lower Precedence values or with null Precedence values.

Two groups can have the same Precedence value. If this happens, neither group takes precedence over the other. If two groups with the same Precedence have the same role ARN, that role is used in the cognito:preferred_role claim in tokens for users in each group. If the two groups have different role ARNs, the cognito:preferred_role claim is not set in users' tokens.

The default Precedence value is null.

groupType_lastModifiedDate :: Lens' GroupType (Maybe UTCTime)

The date the group was last modified.

groupType_roleArn :: Lens' GroupType (Maybe Text)

The role ARN for the group.

groupType_userPoolId :: Lens' GroupType (Maybe Text)

The user pool ID for the user pool.

groupType_groupName :: Lens' GroupType (Maybe Text)

The name of the group.

groupType_creationDate :: Lens' GroupType (Maybe UTCTime)

The date the group was created.

groupType_description :: Lens' GroupType (Maybe Text)

A string containing the description of the group.

groupType_precedence :: Lens' GroupType (Maybe Natural)

A nonnegative integer value that specifies the precedence of this group relative to the other groups that a user can belong to in the user pool. If a user belongs to two or more groups, it is the group with the highest precedence whose role ARN will be used in the cognito:roles and cognito:preferred_role claims in the user's tokens. Groups with higher Precedence values take precedence over groups with lower Precedence values or with null Precedence values.

Two groups can have the same Precedence value. If this happens, neither group takes precedence over the other. If two groups with the same Precedence have the same role ARN, that role is used in the cognito:preferred_role claim in tokens for users in each group. If the two groups have different role ARNs, the cognito:preferred_role claim is not set in users' tokens.

The default Precedence value is null.