Network.AWS.CognitoIdentityProvider.Types.IdentityProviderType

Description

Documentation

data IdentityProviderType

A container for information about an identity provider.

See: newIdentityProviderType smart constructor.

Constructors

IdentityProviderType'

Fields

lastModifiedDate :: Maybe POSIX

The date the identity provider was last modified.

providerType :: Maybe IdentityProviderTypeType

The identity provider type.

providerName :: Maybe Text

The identity provider name.

providerDetails :: Maybe (HashMap Text Text)

The identity provider details. The following list describes the provider detail keys for each identity provider type.

For Google and Login with Amazon:

client_id

client_secret

authorize_scopes

For Facebook:

client_id

client_secret

authorize_scopes

api_version

For Sign in with Apple:

client_id

team_id

key_id

private_key

authorize_scopes

For OIDC providers:

client_id

client_secret

attributes_request_method

oidc_issuer

authorize_scopes

authorize_url /if not available from discovery URL specified by oidc_issuer key/

token_url /if not available from discovery URL specified by oidc_issuer key/

attributes_url /if not available from discovery URL specified by oidc_issuer key/

jwks_uri /if not available from discovery URL specified by oidc_issuer key/

For SAML providers:

MetadataFile OR MetadataURL

IDPSignOut optional

userPoolId :: Maybe Text

The user pool ID.

creationDate :: Maybe POSIX

The date the identity provider was created.

idpIdentifiers :: Maybe [Text]

A list of identity provider identifiers.

attributeMapping :: Maybe (HashMap Text Text)

A mapping of identity provider attributes to standard and custom user pool attributes.

newIdentityProviderType :: IdentityProviderType

Create a value of IdentityProviderType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedDate:IdentityProviderType', identityProviderType_lastModifiedDate - The date the identity provider was last modified.

$sel:providerType:IdentityProviderType', identityProviderType_providerType - The identity provider type.

$sel:providerName:IdentityProviderType', identityProviderType_providerName - The identity provider name.

$sel:providerDetails:IdentityProviderType', identityProviderType_providerDetails - The identity provider details. The following list describes the provider detail keys for each identity provider type.

For Google and Login with Amazon:

client_id

client_secret

authorize_scopes

For Facebook:

client_id

client_secret

authorize_scopes

api_version

For Sign in with Apple:

client_id

team_id

key_id

private_key

authorize_scopes

For OIDC providers:

client_id

client_secret

attributes_request_method

oidc_issuer

authorize_scopes

authorize_url /if not available from discovery URL specified by oidc_issuer key/

token_url /if not available from discovery URL specified by oidc_issuer key/

attributes_url /if not available from discovery URL specified by oidc_issuer key/

jwks_uri /if not available from discovery URL specified by oidc_issuer key/

For SAML providers:

MetadataFile OR MetadataURL

IDPSignOut optional

$sel:userPoolId:IdentityProviderType', identityProviderType_userPoolId - The user pool ID.

$sel:creationDate:IdentityProviderType', identityProviderType_creationDate - The date the identity provider was created.

$sel:idpIdentifiers:IdentityProviderType', identityProviderType_idpIdentifiers - A list of identity provider identifiers.

$sel:attributeMapping:IdentityProviderType', identityProviderType_attributeMapping - A mapping of identity provider attributes to standard and custom user pool attributes.

identityProviderType_lastModifiedDate :: Lens' IdentityProviderType (Maybe UTCTime)

The date the identity provider was last modified.

identityProviderType_providerType :: Lens' IdentityProviderType (Maybe IdentityProviderTypeType)

The identity provider type.

identityProviderType_providerName :: Lens' IdentityProviderType (Maybe Text)

The identity provider name.

identityProviderType_providerDetails :: Lens' IdentityProviderType (Maybe (HashMap Text Text))

The identity provider details. The following list describes the provider detail keys for each identity provider type.

For Google and Login with Amazon:

client_id

client_secret

authorize_scopes

For Facebook:

client_id

client_secret

authorize_scopes

api_version

For Sign in with Apple:

client_id

team_id

key_id

private_key

authorize_scopes

For OIDC providers:

client_id

client_secret

attributes_request_method

oidc_issuer

authorize_scopes

authorize_url /if not available from discovery URL specified by oidc_issuer key/

token_url /if not available from discovery URL specified by oidc_issuer key/

attributes_url /if not available from discovery URL specified by oidc_issuer key/

jwks_uri /if not available from discovery URL specified by oidc_issuer key/

For SAML providers:

MetadataFile OR MetadataURL

IDPSignOut optional

identityProviderType_userPoolId :: Lens' IdentityProviderType (Maybe Text)

The user pool ID.

identityProviderType_creationDate :: Lens' IdentityProviderType (Maybe UTCTime)

The date the identity provider was created.

identityProviderType_idpIdentifiers :: Lens' IdentityProviderType (Maybe [Text])

A list of identity provider identifiers.

identityProviderType_attributeMapping :: Lens' IdentityProviderType (Maybe (HashMap Text Text))

A mapping of identity provider attributes to standard and custom user pool attributes.