Network.AWS.CognitoIdentityProvider.Types.LambdaConfigType

Description

Documentation

data LambdaConfigType

Specifies the configuration for Lambda triggers.

See: newLambdaConfigType smart constructor.

Constructors

LambdaConfigType'

Fields

preAuthentication :: Maybe Text

A pre-authentication Lambda trigger.

createAuthChallenge :: Maybe Text

Creates an authentication challenge.

verifyAuthChallengeResponse :: Maybe Text

Verifies the authentication challenge response.

customSMSSender :: Maybe CustomSMSLambdaVersionConfigType

A custom SMS sender Lambda trigger.

postAuthentication :: Maybe Text

A post-authentication Lambda trigger.

customMessage :: Maybe Text

A custom Message Lambda trigger.

defineAuthChallenge :: Maybe Text

Defines the authentication challenge.

customEmailSender :: Maybe CustomEmailLambdaVersionConfigType

A custom email sender Lambda trigger.

kmsKeyID :: Maybe Text

The Amazon Resource Name of Key Management Service Customer master keys . Amazon Cognito uses the key to encrypt codes and temporary passwords sent to CustomEmailSender and CustomSMSSender.

postConfirmation :: Maybe Text

A post-confirmation Lambda trigger.

preTokenGeneration :: Maybe Text

A Lambda trigger that is invoked before token generation.

userMigration :: Maybe Text

The user migration Lambda config type.

preSignUp :: Maybe Text

A pre-registration Lambda trigger.

newLambdaConfigType :: LambdaConfigType

Create a value of LambdaConfigType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:preAuthentication:LambdaConfigType', lambdaConfigType_preAuthentication - A pre-authentication Lambda trigger.

$sel:createAuthChallenge:LambdaConfigType', lambdaConfigType_createAuthChallenge - Creates an authentication challenge.

$sel:verifyAuthChallengeResponse:LambdaConfigType', lambdaConfigType_verifyAuthChallengeResponse - Verifies the authentication challenge response.

$sel:customSMSSender:LambdaConfigType', lambdaConfigType_customSMSSender - A custom SMS sender Lambda trigger.

$sel:postAuthentication:LambdaConfigType', lambdaConfigType_postAuthentication - A post-authentication Lambda trigger.

$sel:customMessage:LambdaConfigType', lambdaConfigType_customMessage - A custom Message Lambda trigger.

$sel:defineAuthChallenge:LambdaConfigType', lambdaConfigType_defineAuthChallenge - Defines the authentication challenge.

$sel:customEmailSender:LambdaConfigType', lambdaConfigType_customEmailSender - A custom email sender Lambda trigger.

$sel:kmsKeyID:LambdaConfigType', lambdaConfigType_kmsKeyID - The Amazon Resource Name of Key Management Service Customer master keys . Amazon Cognito uses the key to encrypt codes and temporary passwords sent to CustomEmailSender and CustomSMSSender.

$sel:postConfirmation:LambdaConfigType', lambdaConfigType_postConfirmation - A post-confirmation Lambda trigger.

$sel:preTokenGeneration:LambdaConfigType', lambdaConfigType_preTokenGeneration - A Lambda trigger that is invoked before token generation.

$sel:userMigration:LambdaConfigType', lambdaConfigType_userMigration - The user migration Lambda config type.

$sel:preSignUp:LambdaConfigType', lambdaConfigType_preSignUp - A pre-registration Lambda trigger.

lambdaConfigType_preAuthentication :: Lens' LambdaConfigType (Maybe Text)

A pre-authentication Lambda trigger.

lambdaConfigType_createAuthChallenge :: Lens' LambdaConfigType (Maybe Text)

Creates an authentication challenge.

lambdaConfigType_verifyAuthChallengeResponse :: Lens' LambdaConfigType (Maybe Text)

Verifies the authentication challenge response.

lambdaConfigType_customSMSSender :: Lens' LambdaConfigType (Maybe CustomSMSLambdaVersionConfigType)

A custom SMS sender Lambda trigger.

lambdaConfigType_postAuthentication :: Lens' LambdaConfigType (Maybe Text)

A post-authentication Lambda trigger.

lambdaConfigType_customMessage :: Lens' LambdaConfigType (Maybe Text)

A custom Message Lambda trigger.

lambdaConfigType_defineAuthChallenge :: Lens' LambdaConfigType (Maybe Text)

Defines the authentication challenge.

lambdaConfigType_customEmailSender :: Lens' LambdaConfigType (Maybe CustomEmailLambdaVersionConfigType)

A custom email sender Lambda trigger.

lambdaConfigType_kmsKeyID :: Lens' LambdaConfigType (Maybe Text)

The Amazon Resource Name of Key Management Service Customer master keys . Amazon Cognito uses the key to encrypt codes and temporary passwords sent to CustomEmailSender and CustomSMSSender.

lambdaConfigType_postConfirmation :: Lens' LambdaConfigType (Maybe Text)

A post-confirmation Lambda trigger.

lambdaConfigType_preTokenGeneration :: Lens' LambdaConfigType (Maybe Text)

A Lambda trigger that is invoked before token generation.

lambdaConfigType_userMigration :: Lens' LambdaConfigType (Maybe Text)

The user migration Lambda config type.

lambdaConfigType_preSignUp :: Lens' LambdaConfigType (Maybe Text)

A pre-registration Lambda trigger.