Network.AWS.CognitoIdentityProvider.Types.PasswordPolicyType

Description

Documentation

data PasswordPolicyType

The password policy type.

See: newPasswordPolicyType smart constructor.

Constructors

PasswordPolicyType'

Fields

temporaryPasswordValidityDays :: Maybe Natural

In the password policy you have set, refers to the number of days a temporary password is valid. If the user does not sign-in during this time, their password will need to be reset by an administrator.

When you set TemporaryPasswordValidityDays for a user pool, you will no longer be able to set the deprecated UnusedAccountValidityDays value for that user pool.

requireUppercase :: Maybe Bool

In the password policy that you have set, refers to whether you have required users to use at least one uppercase letter in their password.

minimumLength :: Maybe Natural

The minimum length of the password policy that you have set. Cannot be less than 6.

requireSymbols :: Maybe Bool

In the password policy that you have set, refers to whether you have required users to use at least one symbol in their password.

requireNumbers :: Maybe Bool

In the password policy that you have set, refers to whether you have required users to use at least one number in their password.

requireLowercase :: Maybe Bool

In the password policy that you have set, refers to whether you have required users to use at least one lowercase letter in their password.

newPasswordPolicyType :: PasswordPolicyType

Create a value of PasswordPolicyType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:temporaryPasswordValidityDays:PasswordPolicyType', passwordPolicyType_temporaryPasswordValidityDays - In the password policy you have set, refers to the number of days a temporary password is valid. If the user does not sign-in during this time, their password will need to be reset by an administrator.

When you set TemporaryPasswordValidityDays for a user pool, you will no longer be able to set the deprecated UnusedAccountValidityDays value for that user pool.

$sel:requireUppercase:PasswordPolicyType', passwordPolicyType_requireUppercase - In the password policy that you have set, refers to whether you have required users to use at least one uppercase letter in their password.

$sel:minimumLength:PasswordPolicyType', passwordPolicyType_minimumLength - The minimum length of the password policy that you have set. Cannot be less than 6.

$sel:requireSymbols:PasswordPolicyType', passwordPolicyType_requireSymbols - In the password policy that you have set, refers to whether you have required users to use at least one symbol in their password.

$sel:requireNumbers:PasswordPolicyType', passwordPolicyType_requireNumbers - In the password policy that you have set, refers to whether you have required users to use at least one number in their password.

$sel:requireLowercase:PasswordPolicyType', passwordPolicyType_requireLowercase - In the password policy that you have set, refers to whether you have required users to use at least one lowercase letter in their password.

passwordPolicyType_temporaryPasswordValidityDays :: Lens' PasswordPolicyType (Maybe Natural)

In the password policy you have set, refers to the number of days a temporary password is valid. If the user does not sign-in during this time, their password will need to be reset by an administrator.

When you set TemporaryPasswordValidityDays for a user pool, you will no longer be able to set the deprecated UnusedAccountValidityDays value for that user pool.

passwordPolicyType_requireUppercase :: Lens' PasswordPolicyType (Maybe Bool)

In the password policy that you have set, refers to whether you have required users to use at least one uppercase letter in their password.

passwordPolicyType_minimumLength :: Lens' PasswordPolicyType (Maybe Natural)

The minimum length of the password policy that you have set. Cannot be less than 6.

passwordPolicyType_requireSymbols :: Lens' PasswordPolicyType (Maybe Bool)

In the password policy that you have set, refers to whether you have required users to use at least one symbol in their password.

passwordPolicyType_requireNumbers :: Lens' PasswordPolicyType (Maybe Bool)

In the password policy that you have set, refers to whether you have required users to use at least one number in their password.

passwordPolicyType_requireLowercase :: Lens' PasswordPolicyType (Maybe Bool)

In the password policy that you have set, refers to whether you have required users to use at least one lowercase letter in their password.