Network.AWS.CognitoIdentityProvider.Types.RiskExceptionConfigurationType

Description

Documentation

data RiskExceptionConfigurationType

The type of the configuration to override the risk decision.

See: newRiskExceptionConfigurationType smart constructor.

Constructors

RiskExceptionConfigurationType'

Fields

skippedIPRangeList :: Maybe [Text]

Risk detection is not performed on the IP addresses in the range list. The IP range is in CIDR notation.

blockedIPRangeList :: Maybe [Text]

Overrides the risk decision to always block the pre-authentication requests. The IP range is in CIDR notation: a compact representation of an IP address and its associated routing prefix.

newRiskExceptionConfigurationType :: RiskExceptionConfigurationType

Create a value of RiskExceptionConfigurationType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:skippedIPRangeList:RiskExceptionConfigurationType', riskExceptionConfigurationType_skippedIPRangeList - Risk detection is not performed on the IP addresses in the range list. The IP range is in CIDR notation.

$sel:blockedIPRangeList:RiskExceptionConfigurationType', riskExceptionConfigurationType_blockedIPRangeList - Overrides the risk decision to always block the pre-authentication requests. The IP range is in CIDR notation: a compact representation of an IP address and its associated routing prefix.

riskExceptionConfigurationType_skippedIPRangeList :: Lens' RiskExceptionConfigurationType (Maybe [Text])

Risk detection is not performed on the IP addresses in the range list. The IP range is in CIDR notation.

riskExceptionConfigurationType_blockedIPRangeList :: Lens' RiskExceptionConfigurationType (Maybe [Text])

Overrides the risk decision to always block the pre-authentication requests. The IP range is in CIDR notation: a compact representation of an IP address and its associated routing prefix.