Network.AWS.CognitoIdentityProvider.Types.SchemaAttributeType

Description

Documentation

data SchemaAttributeType

Contains information about the schema attribute.

See: newSchemaAttributeType smart constructor.

Constructors

SchemaAttributeType'

Fields

attributeDataType :: Maybe AttributeDataType

The attribute data type.

numberAttributeConstraints :: Maybe NumberAttributeConstraintsType

Specifies the constraints for an attribute of the number type.

required :: Maybe Bool

Specifies whether a user pool attribute is required. If the attribute is required and the user does not provide a value, registration or sign-in will fail.

developerOnlyAttribute :: Maybe Bool

We recommend that you use WriteAttributes in the user pool client to control how attributes can be mutated for new use cases instead of using DeveloperOnlyAttribute.

Specifies whether the attribute type is developer only. This attribute can only be modified by an administrator. Users will not be able to modify this attribute using their access token. For example, DeveloperOnlyAttribute can be modified using AdminUpdateUserAttributes but cannot be updated using UpdateUserAttributes.

name :: Maybe Text

A schema attribute of the name type.

stringAttributeConstraints :: Maybe StringAttributeConstraintsType

Specifies the constraints for an attribute of the string type.

mutable :: Maybe Bool

Specifies whether the value of the attribute can be changed.

For any user pool attribute that's mapped to an identity provider attribute, you must set this parameter to true. Amazon Cognito updates mapped attributes when users sign in to your application through an identity provider. If an attribute is immutable, Amazon Cognito throws an error when it attempts to update the attribute. For more information, see Specifying Identity Provider Attribute Mappings for Your User Pool.

newSchemaAttributeType :: SchemaAttributeType

Create a value of SchemaAttributeType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributeDataType:SchemaAttributeType', schemaAttributeType_attributeDataType - The attribute data type.

$sel:numberAttributeConstraints:SchemaAttributeType', schemaAttributeType_numberAttributeConstraints - Specifies the constraints for an attribute of the number type.

$sel:required:SchemaAttributeType', schemaAttributeType_required - Specifies whether a user pool attribute is required. If the attribute is required and the user does not provide a value, registration or sign-in will fail.

$sel:developerOnlyAttribute:SchemaAttributeType', schemaAttributeType_developerOnlyAttribute - We recommend that you use WriteAttributes in the user pool client to control how attributes can be mutated for new use cases instead of using DeveloperOnlyAttribute.

Specifies whether the attribute type is developer only. This attribute can only be modified by an administrator. Users will not be able to modify this attribute using their access token. For example, DeveloperOnlyAttribute can be modified using AdminUpdateUserAttributes but cannot be updated using UpdateUserAttributes.

$sel:name:SchemaAttributeType', schemaAttributeType_name - A schema attribute of the name type.

$sel:stringAttributeConstraints:SchemaAttributeType', schemaAttributeType_stringAttributeConstraints - Specifies the constraints for an attribute of the string type.

$sel:mutable:SchemaAttributeType', schemaAttributeType_mutable - Specifies whether the value of the attribute can be changed.

For any user pool attribute that's mapped to an identity provider attribute, you must set this parameter to true. Amazon Cognito updates mapped attributes when users sign in to your application through an identity provider. If an attribute is immutable, Amazon Cognito throws an error when it attempts to update the attribute. For more information, see Specifying Identity Provider Attribute Mappings for Your User Pool.

schemaAttributeType_attributeDataType :: Lens' SchemaAttributeType (Maybe AttributeDataType)

The attribute data type.

schemaAttributeType_numberAttributeConstraints :: Lens' SchemaAttributeType (Maybe NumberAttributeConstraintsType)

Specifies the constraints for an attribute of the number type.

schemaAttributeType_required :: Lens' SchemaAttributeType (Maybe Bool)

Specifies whether a user pool attribute is required. If the attribute is required and the user does not provide a value, registration or sign-in will fail.

schemaAttributeType_developerOnlyAttribute :: Lens' SchemaAttributeType (Maybe Bool)

We recommend that you use WriteAttributes in the user pool client to control how attributes can be mutated for new use cases instead of using DeveloperOnlyAttribute.

Specifies whether the attribute type is developer only. This attribute can only be modified by an administrator. Users will not be able to modify this attribute using their access token. For example, DeveloperOnlyAttribute can be modified using AdminUpdateUserAttributes but cannot be updated using UpdateUserAttributes.

schemaAttributeType_name :: Lens' SchemaAttributeType (Maybe Text)

A schema attribute of the name type.

schemaAttributeType_stringAttributeConstraints :: Lens' SchemaAttributeType (Maybe StringAttributeConstraintsType)

Specifies the constraints for an attribute of the string type.

schemaAttributeType_mutable :: Lens' SchemaAttributeType (Maybe Bool)

Specifies whether the value of the attribute can be changed.

For any user pool attribute that's mapped to an identity provider attribute, you must set this parameter to true. Amazon Cognito updates mapped attributes when users sign in to your application through an identity provider. If an attribute is immutable, Amazon Cognito throws an error when it attempts to update the attribute. For more information, see Specifying Identity Provider Attribute Mappings for Your User Pool.