Network.AWS.CognitoIdentityProvider.Types.TokenValidityUnitsType

Description

Documentation

data TokenValidityUnitsType

The data type for TokenValidityUnits that specifics the time measurements for token validity.

See: newTokenValidityUnitsType smart constructor.

Constructors

TokenValidityUnitsType'

Fields

accessToken :: Maybe TimeUnitsType

A time unit in “seconds”, “minutes”, “hours” or “days” for the value in AccessTokenValidity, defaults to hours.

refreshToken :: Maybe TimeUnitsType

A time unit in “seconds”, “minutes”, “hours” or “days” for the value in RefreshTokenValidity, defaults to days.

idToken :: Maybe TimeUnitsType

A time unit in “seconds”, “minutes”, “hours” or “days” for the value in IdTokenValidity, defaults to hours.

newTokenValidityUnitsType :: TokenValidityUnitsType

Create a value of TokenValidityUnitsType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accessToken:TokenValidityUnitsType', tokenValidityUnitsType_accessToken - A time unit in “seconds”, “minutes”, “hours” or “days” for the value in AccessTokenValidity, defaults to hours.

$sel:refreshToken:TokenValidityUnitsType', tokenValidityUnitsType_refreshToken - A time unit in “seconds”, “minutes”, “hours” or “days” for the value in RefreshTokenValidity, defaults to days.

$sel:idToken:TokenValidityUnitsType', tokenValidityUnitsType_idToken - A time unit in “seconds”, “minutes”, “hours” or “days” for the value in IdTokenValidity, defaults to hours.

tokenValidityUnitsType_accessToken :: Lens' TokenValidityUnitsType (Maybe TimeUnitsType)

A time unit in “seconds”, “minutes”, “hours” or “days” for the value in AccessTokenValidity, defaults to hours.

tokenValidityUnitsType_refreshToken :: Lens' TokenValidityUnitsType (Maybe TimeUnitsType)

A time unit in “seconds”, “minutes”, “hours” or “days” for the value in RefreshTokenValidity, defaults to days.

tokenValidityUnitsType_idToken :: Lens' TokenValidityUnitsType (Maybe TimeUnitsType)

A time unit in “seconds”, “minutes”, “hours” or “days” for the value in IdTokenValidity, defaults to hours.