Network.AWS.CognitoIdentityProvider.Types.UserPoolClientDescription

Description

Documentation

data UserPoolClientDescription

The description of the user pool client.

See: newUserPoolClientDescription smart constructor.

Constructors

UserPoolClientDescription'

Fields

clientId :: Maybe (Sensitive Text)

The ID of the client associated with the user pool.

clientName :: Maybe Text

The client name from the user pool client description.

userPoolId :: Maybe Text

The user pool ID for the user pool where you want to describe the user pool client.

newUserPoolClientDescription :: UserPoolClientDescription

Create a value of UserPoolClientDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientId:UserPoolClientDescription', userPoolClientDescription_clientId - The ID of the client associated with the user pool.

$sel:clientName:UserPoolClientDescription', userPoolClientDescription_clientName - The client name from the user pool client description.

$sel:userPoolId:UserPoolClientDescription', userPoolClientDescription_userPoolId - The user pool ID for the user pool where you want to describe the user pool client.

userPoolClientDescription_clientId :: Lens' UserPoolClientDescription (Maybe Text)

The ID of the client associated with the user pool.

userPoolClientDescription_clientName :: Lens' UserPoolClientDescription (Maybe Text)

The client name from the user pool client description.

userPoolClientDescription_userPoolId :: Lens' UserPoolClientDescription (Maybe Text)

The user pool ID for the user pool where you want to describe the user pool client.