Network.AWS.CognitoIdentityProvider.Types.UserPoolClientType

Description

Documentation

data UserPoolClientType

Contains information about a user pool client.

See: newUserPoolClientType smart constructor.

Constructors

UserPoolClientType'

Fields

lastModifiedDate :: Maybe POSIX

The date the user pool client was last modified.

clientSecret :: Maybe (Sensitive Text)

The client secret from the user pool request of the client type.

idTokenValidity :: Maybe Natural

The time limit, specified by tokenValidityUnits, defaulting to hours, after which the refresh token is no longer valid and cannot be used.

refreshTokenValidity :: Maybe Natural

The time limit, in days, after which the refresh token is no longer valid and cannot be used.

clientId :: Maybe (Sensitive Text)

The ID of the client associated with the user pool.

clientName :: Maybe Text

The client name from the user pool request of the client type.

allowedOAuthScopes :: Maybe [Text]

The allowed OAuth scopes. Possible values provided by OAuth are: phone, email, openid, and profile. Possible values provided by Amazon Web Services are: aws.cognito.signin.user.admin. Custom scopes created in Resource Servers are also supported.

analyticsConfiguration :: Maybe AnalyticsConfigurationType

The Amazon Pinpoint analytics configuration for the user pool client.

Cognito User Pools only supports sending events to Amazon Pinpoint projects in the US East (N. Virginia) us-east-1 Region, regardless of the region in which the user pool resides.

userPoolId :: Maybe Text

The user pool ID for the user pool client.

creationDate :: Maybe POSIX

The date the user pool client was created.

readAttributes :: Maybe [Text]

The Read-only attributes.

supportedIdentityProviders :: Maybe [Text]

A list of provider names for the identity providers that are supported on this client.

writeAttributes :: Maybe [Text]

The writeable attributes.

logoutURLs :: Maybe [Text]

A list of allowed logout URLs for the identity providers.

enableTokenRevocation :: Maybe Bool

Indicates whether token revocation is enabled for the user pool client. When you create a new user pool client, token revocation is enabled by default. For more information about revoking tokens, see RevokeToken.

defaultRedirectURI :: Maybe Text

The default redirect URI. Must be in the CallbackURLs list.

A redirect URI must:

Be an absolute URI.

Be registered with the authorization server.

Not include a fragment component.

See OAuth 2.0 - Redirection Endpoint.

Amazon Cognito requires HTTPS over HTTP except for http://localhost for testing purposes only.

App callback URLs such as myapp://example are also supported.

explicitAuthFlows :: Maybe [ExplicitAuthFlowsType]

The authentication flows that are supported by the user pool clients. Flow names without the ALLOW_ prefix are deprecated in favor of new names with the ALLOW_ prefix. Note that values with ALLOW_ prefix cannot be used along with values without ALLOW_ prefix.

Valid values include:

ALLOW_ADMIN_USER_PASSWORD_AUTH: Enable admin based user password authentication flow ADMIN_USER_PASSWORD_AUTH. This setting replaces the ADMIN_NO_SRP_AUTH setting. With this authentication flow, Cognito receives the password in the request instead of using the SRP (Secure Remote Password protocol) protocol to verify passwords.

ALLOW_CUSTOM_AUTH: Enable Lambda trigger based authentication.

ALLOW_USER_PASSWORD_AUTH: Enable user password-based authentication. In this flow, Cognito receives the password in the request instead of using the SRP protocol to verify passwords.

ALLOW_USER_SRP_AUTH: Enable SRP based authentication.

ALLOW_REFRESH_TOKEN_AUTH: Enable authflow to refresh tokens.

tokenValidityUnits :: Maybe TokenValidityUnitsType

The time units used to specify the token validity times of their respective token.

callbackURLs :: Maybe [Text]

A list of allowed redirect (callback) URLs for the identity providers.

A redirect URI must:

Be an absolute URI.

Be registered with the authorization server.

Not include a fragment component.

See OAuth 2.0 - Redirection Endpoint.

Amazon Cognito requires HTTPS over HTTP except for http://localhost for testing purposes only.

App callback URLs such as myapp://example are also supported.

allowedOAuthFlows :: Maybe [OAuthFlowType]

The allowed OAuth flows.

Set to code to initiate a code grant flow, which provides an authorization code as the response. This code can be exchanged for access tokens with the token endpoint.

Set to implicit to specify that the client should get the access token (and, optionally, ID token, based on scopes) directly.

Set to client_credentials to specify that the client should get the access token (and, optionally, ID token, based on scopes) from the token endpoint using a combination of client and client_secret.

accessTokenValidity :: Maybe Natural

The time limit, specified by tokenValidityUnits, defaulting to hours, after which the access token is no longer valid and cannot be used.

allowedOAuthFlowsUserPoolClient :: Maybe Bool

Set to true if the client is allowed to follow the OAuth protocol when interacting with Cognito user pools.

preventUserExistenceErrors :: Maybe PreventUserExistenceErrorTypes

Use this setting to choose which errors and responses are returned by Cognito APIs during authentication, account confirmation, and password recovery when the user does not exist in the user pool. When set to ENABLED and the user does not exist, authentication returns an error indicating either the username or password was incorrect, and account confirmation and password recovery return a response indicating a code was sent to a simulated destination. When set to LEGACY, those APIs will return a UserNotFoundException exception if the user does not exist in the user pool.

Valid values include:

ENABLED - This prevents user existence-related errors.

LEGACY - This represents the old behavior of Cognito where user existence related errors are not prevented.

After February 15th 2020, the value of PreventUserExistenceErrors will default to ENABLED for newly created user pool clients if no value is provided.

newUserPoolClientType :: UserPoolClientType

Create a value of UserPoolClientType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedDate:UserPoolClientType', userPoolClientType_lastModifiedDate - The date the user pool client was last modified.

$sel:clientSecret:UserPoolClientType', userPoolClientType_clientSecret - The client secret from the user pool request of the client type.

$sel:idTokenValidity:UserPoolClientType', userPoolClientType_idTokenValidity - The time limit, specified by tokenValidityUnits, defaulting to hours, after which the refresh token is no longer valid and cannot be used.

$sel:refreshTokenValidity:UserPoolClientType', userPoolClientType_refreshTokenValidity - The time limit, in days, after which the refresh token is no longer valid and cannot be used.

$sel:clientId:UserPoolClientType', userPoolClientType_clientId - The ID of the client associated with the user pool.

$sel:clientName:UserPoolClientType', userPoolClientType_clientName - The client name from the user pool request of the client type.

$sel:allowedOAuthScopes:UserPoolClientType', userPoolClientType_allowedOAuthScopes - The allowed OAuth scopes. Possible values provided by OAuth are: phone, email, openid, and profile. Possible values provided by Amazon Web Services are: aws.cognito.signin.user.admin. Custom scopes created in Resource Servers are also supported.

$sel:analyticsConfiguration:UserPoolClientType', userPoolClientType_analyticsConfiguration - The Amazon Pinpoint analytics configuration for the user pool client.

Cognito User Pools only supports sending events to Amazon Pinpoint projects in the US East (N. Virginia) us-east-1 Region, regardless of the region in which the user pool resides.

$sel:userPoolId:UserPoolClientType', userPoolClientType_userPoolId - The user pool ID for the user pool client.

$sel:creationDate:UserPoolClientType', userPoolClientType_creationDate - The date the user pool client was created.

$sel:readAttributes:UserPoolClientType', userPoolClientType_readAttributes - The Read-only attributes.

$sel:supportedIdentityProviders:UserPoolClientType', userPoolClientType_supportedIdentityProviders - A list of provider names for the identity providers that are supported on this client.

$sel:writeAttributes:UserPoolClientType', userPoolClientType_writeAttributes - The writeable attributes.

$sel:logoutURLs:UserPoolClientType', userPoolClientType_logoutURLs - A list of allowed logout URLs for the identity providers.

$sel:enableTokenRevocation:UserPoolClientType', userPoolClientType_enableTokenRevocation - Indicates whether token revocation is enabled for the user pool client. When you create a new user pool client, token revocation is enabled by default. For more information about revoking tokens, see RevokeToken.

$sel:defaultRedirectURI:UserPoolClientType', userPoolClientType_defaultRedirectURI - The default redirect URI. Must be in the CallbackURLs list.

A redirect URI must:

Be an absolute URI.

Be registered with the authorization server.

Not include a fragment component.

See OAuth 2.0 - Redirection Endpoint.

Amazon Cognito requires HTTPS over HTTP except for http://localhost for testing purposes only.

App callback URLs such as myapp://example are also supported.

$sel:explicitAuthFlows:UserPoolClientType', userPoolClientType_explicitAuthFlows - The authentication flows that are supported by the user pool clients. Flow names without the ALLOW_ prefix are deprecated in favor of new names with the ALLOW_ prefix. Note that values with ALLOW_ prefix cannot be used along with values without ALLOW_ prefix.

Valid values include:

ALLOW_ADMIN_USER_PASSWORD_AUTH: Enable admin based user password authentication flow ADMIN_USER_PASSWORD_AUTH. This setting replaces the ADMIN_NO_SRP_AUTH setting. With this authentication flow, Cognito receives the password in the request instead of using the SRP (Secure Remote Password protocol) protocol to verify passwords.

ALLOW_CUSTOM_AUTH: Enable Lambda trigger based authentication.

ALLOW_USER_PASSWORD_AUTH: Enable user password-based authentication. In this flow, Cognito receives the password in the request instead of using the SRP protocol to verify passwords.

ALLOW_USER_SRP_AUTH: Enable SRP based authentication.

ALLOW_REFRESH_TOKEN_AUTH: Enable authflow to refresh tokens.

$sel:tokenValidityUnits:UserPoolClientType', userPoolClientType_tokenValidityUnits - The time units used to specify the token validity times of their respective token.

$sel:callbackURLs:UserPoolClientType', userPoolClientType_callbackURLs - A list of allowed redirect (callback) URLs for the identity providers.

A redirect URI must:

Be an absolute URI.

Be registered with the authorization server.

Not include a fragment component.

See OAuth 2.0 - Redirection Endpoint.

Amazon Cognito requires HTTPS over HTTP except for http://localhost for testing purposes only.

App callback URLs such as myapp://example are also supported.

$sel:allowedOAuthFlows:UserPoolClientType', userPoolClientType_allowedOAuthFlows - The allowed OAuth flows.

Set to code to initiate a code grant flow, which provides an authorization code as the response. This code can be exchanged for access tokens with the token endpoint.

Set to implicit to specify that the client should get the access token (and, optionally, ID token, based on scopes) directly.

Set to client_credentials to specify that the client should get the access token (and, optionally, ID token, based on scopes) from the token endpoint using a combination of client and client_secret.

$sel:accessTokenValidity:UserPoolClientType', userPoolClientType_accessTokenValidity - The time limit, specified by tokenValidityUnits, defaulting to hours, after which the access token is no longer valid and cannot be used.

$sel:allowedOAuthFlowsUserPoolClient:UserPoolClientType', userPoolClientType_allowedOAuthFlowsUserPoolClient - Set to true if the client is allowed to follow the OAuth protocol when interacting with Cognito user pools.

$sel:preventUserExistenceErrors:UserPoolClientType', userPoolClientType_preventUserExistenceErrors - Use this setting to choose which errors and responses are returned by Cognito APIs during authentication, account confirmation, and password recovery when the user does not exist in the user pool. When set to ENABLED and the user does not exist, authentication returns an error indicating either the username or password was incorrect, and account confirmation and password recovery return a response indicating a code was sent to a simulated destination. When set to LEGACY, those APIs will return a UserNotFoundException exception if the user does not exist in the user pool.

Valid values include:

ENABLED - This prevents user existence-related errors.

LEGACY - This represents the old behavior of Cognito where user existence related errors are not prevented.

After February 15th 2020, the value of PreventUserExistenceErrors will default to ENABLED for newly created user pool clients if no value is provided.

userPoolClientType_lastModifiedDate :: Lens' UserPoolClientType (Maybe UTCTime)

The date the user pool client was last modified.

userPoolClientType_clientSecret :: Lens' UserPoolClientType (Maybe Text)

The client secret from the user pool request of the client type.

userPoolClientType_idTokenValidity :: Lens' UserPoolClientType (Maybe Natural)

The time limit, specified by tokenValidityUnits, defaulting to hours, after which the refresh token is no longer valid and cannot be used.

userPoolClientType_refreshTokenValidity :: Lens' UserPoolClientType (Maybe Natural)

The time limit, in days, after which the refresh token is no longer valid and cannot be used.

userPoolClientType_clientId :: Lens' UserPoolClientType (Maybe Text)

The ID of the client associated with the user pool.

userPoolClientType_clientName :: Lens' UserPoolClientType (Maybe Text)

The client name from the user pool request of the client type.

userPoolClientType_allowedOAuthScopes :: Lens' UserPoolClientType (Maybe [Text])

The allowed OAuth scopes. Possible values provided by OAuth are: phone, email, openid, and profile. Possible values provided by Amazon Web Services are: aws.cognito.signin.user.admin. Custom scopes created in Resource Servers are also supported.

userPoolClientType_analyticsConfiguration :: Lens' UserPoolClientType (Maybe AnalyticsConfigurationType)

The Amazon Pinpoint analytics configuration for the user pool client.

Cognito User Pools only supports sending events to Amazon Pinpoint projects in the US East (N. Virginia) us-east-1 Region, regardless of the region in which the user pool resides.

userPoolClientType_userPoolId :: Lens' UserPoolClientType (Maybe Text)

The user pool ID for the user pool client.

userPoolClientType_creationDate :: Lens' UserPoolClientType (Maybe UTCTime)

The date the user pool client was created.

userPoolClientType_readAttributes :: Lens' UserPoolClientType (Maybe [Text])

The Read-only attributes.

userPoolClientType_supportedIdentityProviders :: Lens' UserPoolClientType (Maybe [Text])

A list of provider names for the identity providers that are supported on this client.

userPoolClientType_writeAttributes :: Lens' UserPoolClientType (Maybe [Text])

The writeable attributes.

userPoolClientType_logoutURLs :: Lens' UserPoolClientType (Maybe [Text])

A list of allowed logout URLs for the identity providers.

userPoolClientType_enableTokenRevocation :: Lens' UserPoolClientType (Maybe Bool)

Indicates whether token revocation is enabled for the user pool client. When you create a new user pool client, token revocation is enabled by default. For more information about revoking tokens, see RevokeToken.

userPoolClientType_defaultRedirectURI :: Lens' UserPoolClientType (Maybe Text)

The default redirect URI. Must be in the CallbackURLs list.

A redirect URI must:

Be an absolute URI.

Be registered with the authorization server.

Not include a fragment component.

See OAuth 2.0 - Redirection Endpoint.

Amazon Cognito requires HTTPS over HTTP except for http://localhost for testing purposes only.

App callback URLs such as myapp://example are also supported.

userPoolClientType_explicitAuthFlows :: Lens' UserPoolClientType (Maybe [ExplicitAuthFlowsType])

The authentication flows that are supported by the user pool clients. Flow names without the ALLOW_ prefix are deprecated in favor of new names with the ALLOW_ prefix. Note that values with ALLOW_ prefix cannot be used along with values without ALLOW_ prefix.

Valid values include:

ALLOW_ADMIN_USER_PASSWORD_AUTH: Enable admin based user password authentication flow ADMIN_USER_PASSWORD_AUTH. This setting replaces the ADMIN_NO_SRP_AUTH setting. With this authentication flow, Cognito receives the password in the request instead of using the SRP (Secure Remote Password protocol) protocol to verify passwords.

ALLOW_CUSTOM_AUTH: Enable Lambda trigger based authentication.

ALLOW_USER_PASSWORD_AUTH: Enable user password-based authentication. In this flow, Cognito receives the password in the request instead of using the SRP protocol to verify passwords.

ALLOW_USER_SRP_AUTH: Enable SRP based authentication.

ALLOW_REFRESH_TOKEN_AUTH: Enable authflow to refresh tokens.

userPoolClientType_tokenValidityUnits :: Lens' UserPoolClientType (Maybe TokenValidityUnitsType)

The time units used to specify the token validity times of their respective token.

userPoolClientType_callbackURLs :: Lens' UserPoolClientType (Maybe [Text])

A list of allowed redirect (callback) URLs for the identity providers.

A redirect URI must:

Be an absolute URI.

Be registered with the authorization server.

Not include a fragment component.

See OAuth 2.0 - Redirection Endpoint.

Amazon Cognito requires HTTPS over HTTP except for http://localhost for testing purposes only.

App callback URLs such as myapp://example are also supported.

userPoolClientType_allowedOAuthFlows :: Lens' UserPoolClientType (Maybe [OAuthFlowType])

The allowed OAuth flows.

Set to code to initiate a code grant flow, which provides an authorization code as the response. This code can be exchanged for access tokens with the token endpoint.

Set to implicit to specify that the client should get the access token (and, optionally, ID token, based on scopes) directly.

Set to client_credentials to specify that the client should get the access token (and, optionally, ID token, based on scopes) from the token endpoint using a combination of client and client_secret.

userPoolClientType_accessTokenValidity :: Lens' UserPoolClientType (Maybe Natural)

The time limit, specified by tokenValidityUnits, defaulting to hours, after which the access token is no longer valid and cannot be used.

userPoolClientType_allowedOAuthFlowsUserPoolClient :: Lens' UserPoolClientType (Maybe Bool)

Set to true if the client is allowed to follow the OAuth protocol when interacting with Cognito user pools.

userPoolClientType_preventUserExistenceErrors :: Lens' UserPoolClientType (Maybe PreventUserExistenceErrorTypes)

Use this setting to choose which errors and responses are returned by Cognito APIs during authentication, account confirmation, and password recovery when the user does not exist in the user pool. When set to ENABLED and the user does not exist, authentication returns an error indicating either the username or password was incorrect, and account confirmation and password recovery return a response indicating a code was sent to a simulated destination. When set to LEGACY, those APIs will return a UserNotFoundException exception if the user does not exist in the user pool.

Valid values include:

ENABLED - This prevents user existence-related errors.

LEGACY - This represents the old behavior of Cognito where user existence related errors are not prevented.

After February 15th 2020, the value of PreventUserExistenceErrors will default to ENABLED for newly created user pool clients if no value is provided.