Network.AWS.CognitoIdentityProvider.Types.UserPoolDescriptionType

Description

Documentation

data UserPoolDescriptionType

A user pool description.

See: newUserPoolDescriptionType smart constructor.

Constructors

UserPoolDescriptionType'

Fields

status :: Maybe StatusType

The user pool status in a user pool description.

lastModifiedDate :: Maybe POSIX

The date the user pool description was last modified.

name :: Maybe Text

The name in a user pool description.

id :: Maybe Text

The ID in a user pool description.

creationDate :: Maybe POSIX

The date the user pool description was created.

lambdaConfig :: Maybe LambdaConfigType

The Lambda configuration information in a user pool description.

newUserPoolDescriptionType :: UserPoolDescriptionType

Create a value of UserPoolDescriptionType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:UserPoolDescriptionType', userPoolDescriptionType_status - The user pool status in a user pool description.

$sel:lastModifiedDate:UserPoolDescriptionType', userPoolDescriptionType_lastModifiedDate - The date the user pool description was last modified.

$sel:name:UserPoolDescriptionType', userPoolDescriptionType_name - The name in a user pool description.

$sel:id:UserPoolDescriptionType', userPoolDescriptionType_id - The ID in a user pool description.

$sel:creationDate:UserPoolDescriptionType', userPoolDescriptionType_creationDate - The date the user pool description was created.

$sel:lambdaConfig:UserPoolDescriptionType', userPoolDescriptionType_lambdaConfig - The Lambda configuration information in a user pool description.

userPoolDescriptionType_status :: Lens' UserPoolDescriptionType (Maybe StatusType)

The user pool status in a user pool description.

userPoolDescriptionType_lastModifiedDate :: Lens' UserPoolDescriptionType (Maybe UTCTime)

The date the user pool description was last modified.

userPoolDescriptionType_name :: Lens' UserPoolDescriptionType (Maybe Text)

The name in a user pool description.

userPoolDescriptionType_id :: Lens' UserPoolDescriptionType (Maybe Text)

The ID in a user pool description.

userPoolDescriptionType_creationDate :: Lens' UserPoolDescriptionType (Maybe UTCTime)

The date the user pool description was created.

userPoolDescriptionType_lambdaConfig :: Lens' UserPoolDescriptionType (Maybe LambdaConfigType)

The Lambda configuration information in a user pool description.