Network.AWS.CognitoIdentityProvider.Types.UserType

Description

Documentation

data UserType

The user type.

See: newUserType smart constructor.

Constructors

UserType'

Fields

userCreateDate :: Maybe POSIX

The creation date of the user.

userLastModifiedDate :: Maybe POSIX

The last modified date of the user.

enabled :: Maybe Bool

Specifies whether the user is enabled.

attributes :: Maybe [AttributeType]

A container with information about the user type attributes.

username :: Maybe (Sensitive Text)

The user name of the user you wish to describe.

userStatus :: Maybe UserStatusType

The user status. Can be one of the following:

UNCONFIRMED - User has been created but not confirmed.

CONFIRMED - User has been confirmed.

ARCHIVED - User is no longer active.

COMPROMISED - User is disabled due to a potential security threat.

UNKNOWN - User status is not known.

RESET_REQUIRED - User is confirmed, but the user must request a code and reset his or her password before he or she can sign in.

FORCE_CHANGE_PASSWORD - The user is confirmed and the user can sign in using a temporary password, but on first sign-in, the user must change his or her password to a new value before doing anything else.

mfaOptions :: Maybe [MFAOptionType]

The MFA options for the user.

newUserType :: UserType

Create a value of UserType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userCreateDate:UserType', userType_userCreateDate - The creation date of the user.

$sel:userLastModifiedDate:UserType', userType_userLastModifiedDate - The last modified date of the user.

$sel:enabled:UserType', userType_enabled - Specifies whether the user is enabled.

$sel:attributes:UserType', userType_attributes - A container with information about the user type attributes.

$sel:username:UserType', userType_username - The user name of the user you wish to describe.

$sel:userStatus:UserType', userType_userStatus - The user status. Can be one of the following:

UNCONFIRMED - User has been created but not confirmed.

CONFIRMED - User has been confirmed.

ARCHIVED - User is no longer active.

COMPROMISED - User is disabled due to a potential security threat.

UNKNOWN - User status is not known.

RESET_REQUIRED - User is confirmed, but the user must request a code and reset his or her password before he or she can sign in.

FORCE_CHANGE_PASSWORD - The user is confirmed and the user can sign in using a temporary password, but on first sign-in, the user must change his or her password to a new value before doing anything else.

$sel:mfaOptions:UserType', userType_mfaOptions - The MFA options for the user.

userType_userCreateDate :: Lens' UserType (Maybe UTCTime)

The creation date of the user.

userType_userLastModifiedDate :: Lens' UserType (Maybe UTCTime)

The last modified date of the user.

userType_enabled :: Lens' UserType (Maybe Bool)

Specifies whether the user is enabled.

userType_attributes :: Lens' UserType (Maybe [AttributeType])

A container with information about the user type attributes.

userType_username :: Lens' UserType (Maybe Text)

The user name of the user you wish to describe.

userType_userStatus :: Lens' UserType (Maybe UserStatusType)

The user status. Can be one of the following:

UNCONFIRMED - User has been created but not confirmed.

CONFIRMED - User has been confirmed.

ARCHIVED - User is no longer active.

COMPROMISED - User is disabled due to a potential security threat.

UNKNOWN - User status is not known.

RESET_REQUIRED - User is confirmed, but the user must request a code and reset his or her password before he or she can sign in.

FORCE_CHANGE_PASSWORD - The user is confirmed and the user can sign in using a temporary password, but on first sign-in, the user must change his or her password to a new value before doing anything else.

userType_mfaOptions :: Lens' UserType (Maybe [MFAOptionType])

The MFA options for the user.