Network.AWS.CognitoIdentityProvider.Types.VerificationMessageTemplateType

Description

Documentation

data VerificationMessageTemplateType

The template for verification messages.

See: newVerificationMessageTemplateType smart constructor.

Constructors

VerificationMessageTemplateType'

Fields

defaultEmailOption :: Maybe DefaultEmailOptionType

The default email option.

emailSubject :: Maybe Text

The subject line for the email message template. EmailSubject is allowed only if EmailSendingAccount is DEVELOPER.

emailSubjectByLink :: Maybe Text

The subject line for the email message template for sending a confirmation link to the user. EmailSubjectByLink is allowed only EmailSendingAccount is DEVELOPER.

smsMessage :: Maybe Text

The SMS message template.

emailMessageByLink :: Maybe Text

The email message template for sending a confirmation link to the user. EmailMessageByLink is allowed only if EmailSendingAccount is DEVELOPER.

emailMessage :: Maybe Text

The email message template. EmailMessage is allowed only if EmailSendingAccount is DEVELOPER.

newVerificationMessageTemplateType :: VerificationMessageTemplateType

Create a value of VerificationMessageTemplateType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:defaultEmailOption:VerificationMessageTemplateType', verificationMessageTemplateType_defaultEmailOption - The default email option.

$sel:emailSubject:VerificationMessageTemplateType', verificationMessageTemplateType_emailSubject - The subject line for the email message template. EmailSubject is allowed only if EmailSendingAccount is DEVELOPER.

$sel:emailSubjectByLink:VerificationMessageTemplateType', verificationMessageTemplateType_emailSubjectByLink - The subject line for the email message template for sending a confirmation link to the user. EmailSubjectByLink is allowed only EmailSendingAccount is DEVELOPER.

$sel:smsMessage:VerificationMessageTemplateType', verificationMessageTemplateType_smsMessage - The SMS message template.

$sel:emailMessageByLink:VerificationMessageTemplateType', verificationMessageTemplateType_emailMessageByLink - The email message template for sending a confirmation link to the user. EmailMessageByLink is allowed only if EmailSendingAccount is DEVELOPER.

$sel:emailMessage:VerificationMessageTemplateType', verificationMessageTemplateType_emailMessage - The email message template. EmailMessage is allowed only if EmailSendingAccount is DEVELOPER.

verificationMessageTemplateType_defaultEmailOption :: Lens' VerificationMessageTemplateType (Maybe DefaultEmailOptionType)

The default email option.

verificationMessageTemplateType_emailSubject :: Lens' VerificationMessageTemplateType (Maybe Text)

The subject line for the email message template. EmailSubject is allowed only if EmailSendingAccount is DEVELOPER.

verificationMessageTemplateType_emailSubjectByLink :: Lens' VerificationMessageTemplateType (Maybe Text)

The subject line for the email message template for sending a confirmation link to the user. EmailSubjectByLink is allowed only EmailSendingAccount is DEVELOPER.

verificationMessageTemplateType_smsMessage :: Lens' VerificationMessageTemplateType (Maybe Text)

The SMS message template.

verificationMessageTemplateType_emailMessageByLink :: Lens' VerificationMessageTemplateType (Maybe Text)

The email message template for sending a confirmation link to the user. EmailMessageByLink is allowed only if EmailSendingAccount is DEVELOPER.

verificationMessageTemplateType_emailMessage :: Lens' VerificationMessageTemplateType (Maybe Text)

The email message template. EmailMessage is allowed only if EmailSendingAccount is DEVELOPER.