Network.AWS.CognitoIdentityProvider.UpdateDeviceStatus

Description

Updates the device status.

Creating a Request

data UpdateDeviceStatus

Represents the request to update the device status.

See: newUpdateDeviceStatus smart constructor.

Constructors

UpdateDeviceStatus'

Fields

deviceRememberedStatus :: Maybe DeviceRememberedStatusType

The status of whether a device is remembered.

accessToken :: Sensitive Text

The access token.

deviceKey :: Text

The device key.

newUpdateDeviceStatus

Arguments

:: Text
$sel:accessToken:UpdateDeviceStatus'

-> Text
$sel:deviceKey:UpdateDeviceStatus'

-> UpdateDeviceStatus

Create a value of UpdateDeviceStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceRememberedStatus:UpdateDeviceStatus', updateDeviceStatus_deviceRememberedStatus - The status of whether a device is remembered.

$sel:accessToken:UpdateDeviceStatus', updateDeviceStatus_accessToken - The access token.

$sel:deviceKey:UpdateDeviceStatus', updateDeviceStatus_deviceKey - The device key.

Request Lenses

updateDeviceStatus_deviceRememberedStatus :: Lens' UpdateDeviceStatus (Maybe DeviceRememberedStatusType)

The status of whether a device is remembered.

updateDeviceStatus_accessToken :: Lens' UpdateDeviceStatus Text

The access token.

updateDeviceStatus_deviceKey :: Lens' UpdateDeviceStatus Text

The device key.

Destructuring the Response

data UpdateDeviceStatusResponse

The response to the request to update the device status.

See: newUpdateDeviceStatusResponse smart constructor.

Constructors

UpdateDeviceStatusResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateDeviceStatusResponse

Arguments

:: Int
$sel:httpStatus:UpdateDeviceStatusResponse'

-> UpdateDeviceStatusResponse

Create a value of UpdateDeviceStatusResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateDeviceStatusResponse', updateDeviceStatusResponse_httpStatus - The response's http status code.

Response Lenses

updateDeviceStatusResponse_httpStatus :: Lens' UpdateDeviceStatusResponse Int

The response's http status code.