Network.AWS.CognitoIdentityProvider.UpdateUserPoolDomain

Description

Updates the Secure Sockets Layer (SSL) certificate for the custom domain for your user pool.

You can use this operation to provide the Amazon Resource Name (ARN) of a new certificate to Amazon Cognito. You cannot use it to change the domain for a user pool.

A custom domain is used to host the Amazon Cognito hosted UI, which provides sign-up and sign-in pages for your application. When you set up a custom domain, you provide a certificate that you manage with Certificate Manager (ACM). When necessary, you can use this operation to change the certificate that you applied to your custom domain.

Usually, this is unnecessary following routine certificate renewal with ACM. When you renew your existing certificate in ACM, the ARN for your certificate remains the same, and your custom domain uses the new certificate automatically.

However, if you replace your existing certificate with a new one, ACM gives the new certificate a new ARN. To apply the new certificate to your custom domain, you must provide this ARN to Amazon Cognito.

When you add your new certificate in ACM, you must choose US East (N. Virginia) as the Region.

After you submit your request, Amazon Cognito requires up to 1 hour to distribute your new certificate to your custom domain.

For more information about adding a custom domain to your user pool, see Using Your Own Domain for the Hosted UI.

Creating a Request

data UpdateUserPoolDomain

The UpdateUserPoolDomain request input.

See: newUpdateUserPoolDomain smart constructor.

Constructors

UpdateUserPoolDomain'

Fields

domain :: Text

The domain name for the custom domain that hosts the sign-up and sign-in pages for your application. For example: auth.example.com.

This string can include only lowercase letters, numbers, and hyphens. Do not use a hyphen for the first or last character. Use periods to separate subdomain names.

userPoolId :: Text

The ID of the user pool that is associated with the custom domain that you are updating the certificate for.

customDomainConfig :: CustomDomainConfigType

The configuration for a custom domain that hosts the sign-up and sign-in pages for your application. Use this object to specify an SSL certificate that is managed by ACM.

newUpdateUserPoolDomain

Arguments

:: Text
$sel:domain:UpdateUserPoolDomain'

-> Text
$sel:userPoolId:UpdateUserPoolDomain'

-> CustomDomainConfigType
$sel:customDomainConfig:UpdateUserPoolDomain'

-> UpdateUserPoolDomain

Create a value of UpdateUserPoolDomain with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domain:UpdateUserPoolDomain', updateUserPoolDomain_domain - The domain name for the custom domain that hosts the sign-up and sign-in pages for your application. For example: auth.example.com.

This string can include only lowercase letters, numbers, and hyphens. Do not use a hyphen for the first or last character. Use periods to separate subdomain names.

$sel:userPoolId:UpdateUserPoolDomain', updateUserPoolDomain_userPoolId - The ID of the user pool that is associated with the custom domain that you are updating the certificate for.

$sel:customDomainConfig:UpdateUserPoolDomain', updateUserPoolDomain_customDomainConfig - The configuration for a custom domain that hosts the sign-up and sign-in pages for your application. Use this object to specify an SSL certificate that is managed by ACM.

Request Lenses

updateUserPoolDomain_domain :: Lens' UpdateUserPoolDomain Text

The domain name for the custom domain that hosts the sign-up and sign-in pages for your application. For example: auth.example.com.

This string can include only lowercase letters, numbers, and hyphens. Do not use a hyphen for the first or last character. Use periods to separate subdomain names.

updateUserPoolDomain_userPoolId :: Lens' UpdateUserPoolDomain Text

The ID of the user pool that is associated with the custom domain that you are updating the certificate for.

updateUserPoolDomain_customDomainConfig :: Lens' UpdateUserPoolDomain CustomDomainConfigType

The configuration for a custom domain that hosts the sign-up and sign-in pages for your application. Use this object to specify an SSL certificate that is managed by ACM.

Destructuring the Response

data UpdateUserPoolDomainResponse

The UpdateUserPoolDomain response output.

See: newUpdateUserPoolDomainResponse smart constructor.

Constructors

UpdateUserPoolDomainResponse'

Fields

cloudFrontDomain :: Maybe Text

The Amazon CloudFront endpoint that Amazon Cognito set up when you added the custom domain to your user pool.

httpStatus :: Int

The response's http status code.

newUpdateUserPoolDomainResponse

Arguments

:: Int
$sel:httpStatus:UpdateUserPoolDomainResponse'

-> UpdateUserPoolDomainResponse

Create a value of UpdateUserPoolDomainResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cloudFrontDomain:UpdateUserPoolDomainResponse', updateUserPoolDomainResponse_cloudFrontDomain - The Amazon CloudFront endpoint that Amazon Cognito set up when you added the custom domain to your user pool.

$sel:httpStatus:UpdateUserPoolDomainResponse', updateUserPoolDomainResponse_httpStatus - The response's http status code.

Response Lenses

updateUserPoolDomainResponse_cloudFrontDomain :: Lens' UpdateUserPoolDomainResponse (Maybe Text)

The Amazon CloudFront endpoint that Amazon Cognito set up when you added the custom domain to your user pool.

updateUserPoolDomainResponse_httpStatus :: Lens' UpdateUserPoolDomainResponse Int

The response's http status code.