Network.AWS.CognitoIdentityProvider.VerifySoftwareToken

Description

Use this API to register a user's entered TOTP code and mark the user's software token MFA status as "verified" if successful. The request takes an access token or a session string, but not both.

Creating a Request

data VerifySoftwareToken

See: newVerifySoftwareToken smart constructor.

Constructors

VerifySoftwareToken'

Fields

accessToken :: Maybe (Sensitive Text)

The access token.

friendlyDeviceName :: Maybe Text

The friendly device name.

session :: Maybe Text

The session which should be passed both ways in challenge-response calls to the service.

userCode :: Text

The one time password computed using the secret code returned by AssociateSoftwareToken".

newVerifySoftwareToken

Arguments

:: Text
$sel:userCode:VerifySoftwareToken'

-> VerifySoftwareToken

Create a value of VerifySoftwareToken with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accessToken:VerifySoftwareToken', verifySoftwareToken_accessToken - The access token.

$sel:friendlyDeviceName:VerifySoftwareToken', verifySoftwareToken_friendlyDeviceName - The friendly device name.

$sel:session:VerifySoftwareToken', verifySoftwareToken_session - The session which should be passed both ways in challenge-response calls to the service.

$sel:userCode:VerifySoftwareToken', verifySoftwareToken_userCode - The one time password computed using the secret code returned by AssociateSoftwareToken".

Request Lenses

verifySoftwareToken_accessToken :: Lens' VerifySoftwareToken (Maybe Text)

The access token.

verifySoftwareToken_friendlyDeviceName :: Lens' VerifySoftwareToken (Maybe Text)

The friendly device name.

verifySoftwareToken_session :: Lens' VerifySoftwareToken (Maybe Text)

The session which should be passed both ways in challenge-response calls to the service.

verifySoftwareToken_userCode :: Lens' VerifySoftwareToken Text

The one time password computed using the secret code returned by AssociateSoftwareToken".

Destructuring the Response

data VerifySoftwareTokenResponse

See: newVerifySoftwareTokenResponse smart constructor.

Constructors

VerifySoftwareTokenResponse'

Fields

status :: Maybe VerifySoftwareTokenResponseType

The status of the verify software token.

session :: Maybe Text

The session which should be passed both ways in challenge-response calls to the service.

httpStatus :: Int

The response's http status code.

newVerifySoftwareTokenResponse

Arguments

:: Int
$sel:httpStatus:VerifySoftwareTokenResponse'

-> VerifySoftwareTokenResponse

Create a value of VerifySoftwareTokenResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:VerifySoftwareTokenResponse', verifySoftwareTokenResponse_status - The status of the verify software token.

$sel:session:VerifySoftwareToken', verifySoftwareTokenResponse_session - The session which should be passed both ways in challenge-response calls to the service.

$sel:httpStatus:VerifySoftwareTokenResponse', verifySoftwareTokenResponse_httpStatus - The response's http status code.

Response Lenses

verifySoftwareTokenResponse_status :: Lens' VerifySoftwareTokenResponse (Maybe VerifySoftwareTokenResponseType)

The status of the verify software token.

verifySoftwareTokenResponse_session :: Lens' VerifySoftwareTokenResponse (Maybe Text)

The session which should be passed both ways in challenge-response calls to the service.

verifySoftwareTokenResponse_httpStatus :: Lens' VerifySoftwareTokenResponse Int

The response's http status code.