Network.AWS.CognitoIdentityProvider.VerifyUserAttribute

Description

Verifies the specified user attributes in the user pool.

Creating a Request

data VerifyUserAttribute

Represents the request to verify user attributes.

See: newVerifyUserAttribute smart constructor.

Constructors

VerifyUserAttribute'

Fields

accessToken :: Sensitive Text

Represents the access token of the request to verify user attributes.

attributeName :: Text

The attribute name in the request to verify user attributes.

code :: Text

The verification code in the request to verify user attributes.

newVerifyUserAttribute

Arguments

:: Text
$sel:accessToken:VerifyUserAttribute'

-> Text
$sel:attributeName:VerifyUserAttribute'

-> Text
$sel:code:VerifyUserAttribute'

-> VerifyUserAttribute

Create a value of VerifyUserAttribute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accessToken:VerifyUserAttribute', verifyUserAttribute_accessToken - Represents the access token of the request to verify user attributes.

$sel:attributeName:VerifyUserAttribute', verifyUserAttribute_attributeName - The attribute name in the request to verify user attributes.

$sel:code:VerifyUserAttribute', verifyUserAttribute_code - The verification code in the request to verify user attributes.

Request Lenses

verifyUserAttribute_accessToken :: Lens' VerifyUserAttribute Text

Represents the access token of the request to verify user attributes.

verifyUserAttribute_attributeName :: Lens' VerifyUserAttribute Text

The attribute name in the request to verify user attributes.

verifyUserAttribute_code :: Lens' VerifyUserAttribute Text

The verification code in the request to verify user attributes.

Destructuring the Response

data VerifyUserAttributeResponse

A container representing the response from the server from the request to verify user attributes.

See: newVerifyUserAttributeResponse smart constructor.

Constructors

VerifyUserAttributeResponse'

Fields

httpStatus :: Int

The response's http status code.

newVerifyUserAttributeResponse

Arguments

:: Int
$sel:httpStatus:VerifyUserAttributeResponse'

-> VerifyUserAttributeResponse

Create a value of VerifyUserAttributeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:VerifyUserAttributeResponse', verifyUserAttributeResponse_httpStatus - The response's http status code.

Response Lenses

verifyUserAttributeResponse_httpStatus :: Lens' VerifyUserAttributeResponse Int

The response's http status code.