Network.AWS.CognitoSync

Description

Derived from API version 2014-06-30 of the AWS service descriptions, licensed under Apache 2.0.

Amazon Cognito Sync

Amazon Cognito Sync provides an AWS service and client library that enable cross-device syncing of application-related user data. High-level client libraries are available for both iOS and Android. You can use these libraries to persist data locally so that it's available even if the device is offline. Developer credentials don't need to be stored on the mobile device to access the service. You can use Amazon Cognito to obtain a normalized user ID and credentials. User data is persisted in a dataset that can store up to 1 MB of key-value pairs, and you can have up to 20 datasets per user identity.

With Amazon Cognito Sync, the data stored for each identity is accessible only to credentials assigned to that identity. In order to use the Cognito Sync service, you need to make API calls using credentials retrieved with Amazon Cognito Identity service.

If you want to use Cognito Sync in an Android or iOS application, you will probably want to make API calls via the AWS Mobile SDK. To learn more, see the Developer Guide for Android and the Developer Guide for iOS.

Service Configuration

defaultService :: Service

API version 2014-06-30 of the Amazon Cognito Sync SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by CognitoSync.

InvalidConfigurationException

_InvalidConfigurationException :: AsError a => Getting (First ServiceError) a ServiceError

Prism for InvalidConfigurationException' errors.

ResourceConflictException

_ResourceConflictException :: AsError a => Getting (First ServiceError) a ServiceError

Thrown if an update can't be applied because the resource was changed by another call and this would result in a conflict.

InternalErrorException

_InternalErrorException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates an internal service error.

ConcurrentModificationException

_ConcurrentModificationException :: AsError a => Getting (First ServiceError) a ServiceError

Thrown if there are parallel requests to modify a resource.

InvalidParameterException

_InvalidParameterException :: AsError a => Getting (First ServiceError) a ServiceError

Thrown when a request parameter does not comply with the associated constraints.

AlreadyStreamedException

_AlreadyStreamedException :: AsError a => Getting (First ServiceError) a ServiceError

An exception thrown when a bulk publish operation is requested less than 24 hours after a previous bulk publish operation completed successfully.

InvalidLambdaFunctionOutputException

_InvalidLambdaFunctionOutputException :: AsError a => Getting (First ServiceError) a ServiceError

The AWS Lambda function returned invalid output or an exception.

LambdaThrottledException

_LambdaThrottledException :: AsError a => Getting (First ServiceError) a ServiceError

AWS Lambda throttled your account, please contact AWS Support

LimitExceededException

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Thrown when the limit on the number of objects or operations has been exceeded.

DuplicateRequestException

_DuplicateRequestException :: AsError a => Getting (First ServiceError) a ServiceError

An exception thrown when there is an IN_PROGRESS bulk publish operation for the given identity pool.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

Thrown if the resource doesn't exist.

NotAuthorizedException

_NotAuthorizedException :: AsError a => Getting (First ServiceError) a ServiceError

Thrown when a user is not authorized to access the requested resource.

TooManyRequestsException

_TooManyRequestsException :: AsError a => Getting (First ServiceError) a ServiceError

Thrown if the request is throttled.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

UpdateRecords

data UpdateRecords

A request to post updates to records or add and delete records for a dataset and user.

See: newUpdateRecords smart constructor.

Constructors

UpdateRecords' (Maybe [RecordPatch]) (Maybe Text) (Maybe Text) Text Text Text Text

newUpdateRecords

Arguments

:: Text
$sel:identityPoolId:UpdateRecords'

-> Text
$sel:identityId:UpdateRecords'

-> Text
$sel:datasetName:UpdateRecords'

-> Text
$sel:syncSessionToken:UpdateRecords'

-> UpdateRecords

Create a value of UpdateRecords with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:recordPatches:UpdateRecords', updateRecords_recordPatches - A list of patch operations.

$sel:deviceId:UpdateRecords', updateRecords_deviceId - The unique ID generated for this device by Cognito.

$sel:clientContext:UpdateRecords', updateRecords_clientContext - Intended to supply a device ID that will populate the lastModifiedBy field referenced in other methods. The ClientContext field is not yet implemented.

$sel:identityPoolId:UpdateRecords', updateRecords_identityPoolId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

$sel:identityId:UpdateRecords', updateRecords_identityId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

$sel:datasetName:UpdateRecords', updateRecords_datasetName - A string of up to 128 characters. Allowed characters are a-z, A-Z, 0-9, '_' (underscore), '-' (dash), and '.' (dot).

$sel:syncSessionToken:UpdateRecords', updateRecords_syncSessionToken - The SyncSessionToken returned by a previous call to ListRecords for this dataset and identity.

data UpdateRecordsResponse

Returned for a successful UpdateRecordsRequest.

See: newUpdateRecordsResponse smart constructor.

Constructors

UpdateRecordsResponse' (Maybe [Record]) Int

newUpdateRecordsResponse

Arguments

:: Int
$sel:httpStatus:UpdateRecordsResponse'

-> UpdateRecordsResponse

Create a value of UpdateRecordsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:records:UpdateRecordsResponse', updateRecordsResponse_records - A list of records that have been updated.

$sel:httpStatus:UpdateRecordsResponse', updateRecordsResponse_httpStatus - The response's http status code.

ListRecords

data ListRecords

A request for a list of records.

See: newListRecords smart constructor.

Constructors

ListRecords' (Maybe Text) (Maybe Int) (Maybe Integer) (Maybe Text) Text Text Text

newListRecords

Arguments

:: Text
$sel:identityPoolId:ListRecords'

-> Text
$sel:identityId:ListRecords'

-> Text
$sel:datasetName:ListRecords'

-> ListRecords

Create a value of ListRecords with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListRecords', listRecords_nextToken - A pagination token for obtaining the next page of results.

$sel:maxResults:ListRecords', listRecords_maxResults - The maximum number of results to be returned.

$sel:lastSyncCount:ListRecords', listRecords_lastSyncCount - The last server sync count for this record.

$sel:syncSessionToken:ListRecords', listRecords_syncSessionToken - A token containing a session ID, identity ID, and expiration.

$sel:identityPoolId:ListRecords', listRecords_identityPoolId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

$sel:identityId:ListRecords', listRecords_identityId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

$sel:datasetName:ListRecords', listRecords_datasetName - A string of up to 128 characters. Allowed characters are a-z, A-Z, 0-9, '_' (underscore), '-' (dash), and '.' (dot).

data ListRecordsResponse

Returned for a successful ListRecordsRequest.

See: newListRecordsResponse smart constructor.

Constructors

ListRecordsResponse' (Maybe Text) (Maybe Integer) (Maybe [Record]) (Maybe Bool) (Maybe Int) (Maybe Text) (Maybe Bool) (Maybe Text) (Maybe [Text]) Int

newListRecordsResponse

Arguments

:: Int
$sel:httpStatus:ListRecordsResponse'

-> ListRecordsResponse

Create a value of ListRecordsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListRecords', listRecordsResponse_nextToken - A pagination token for obtaining the next page of results.

$sel:datasetSyncCount:ListRecordsResponse', listRecordsResponse_datasetSyncCount - Server sync count for this dataset.

$sel:records:ListRecordsResponse', listRecordsResponse_records - A list of all records.

$sel:datasetDeletedAfterRequestedSyncCount:ListRecordsResponse', listRecordsResponse_datasetDeletedAfterRequestedSyncCount - A boolean value specifying whether to delete the dataset locally.

$sel:count:ListRecordsResponse', listRecordsResponse_count - Total number of records.

$sel:lastModifiedBy:ListRecordsResponse', listRecordsResponse_lastModifiedBy - The user/device that made the last change to this record.

$sel:datasetExists:ListRecordsResponse', listRecordsResponse_datasetExists - Indicates whether the dataset exists.

$sel:syncSessionToken:ListRecords', listRecordsResponse_syncSessionToken - A token containing a session ID, identity ID, and expiration.

$sel:mergedDatasetNames:ListRecordsResponse', listRecordsResponse_mergedDatasetNames - Names of merged datasets.

$sel:httpStatus:ListRecordsResponse', listRecordsResponse_httpStatus - The response's http status code.

RegisterDevice

data RegisterDevice

A request to RegisterDevice.

See: newRegisterDevice smart constructor.

Constructors

RegisterDevice' Text Text Platform Text

newRegisterDevice

Arguments

:: Text
$sel:identityPoolId:RegisterDevice'

-> Text
$sel:identityId:RegisterDevice'

-> Platform
$sel:platform:RegisterDevice'

-> Text
$sel:token:RegisterDevice'

-> RegisterDevice

Create a value of RegisterDevice with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identityPoolId:RegisterDevice', registerDevice_identityPoolId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. Here, the ID of the pool that the identity belongs to.

$sel:identityId:RegisterDevice', registerDevice_identityId - The unique ID for this identity.

$sel:platform:RegisterDevice', registerDevice_platform - The SNS platform type (e.g. GCM, SDM, APNS, APNS_SANDBOX).

$sel:token:RegisterDevice', registerDevice_token - The push token.

data RegisterDeviceResponse

Response to a RegisterDevice request.

See: newRegisterDeviceResponse smart constructor.

Constructors

RegisterDeviceResponse' (Maybe Text) Int

newRegisterDeviceResponse

Arguments

:: Int
$sel:httpStatus:RegisterDeviceResponse'

-> RegisterDeviceResponse

Create a value of RegisterDeviceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceId:RegisterDeviceResponse', registerDeviceResponse_deviceId - The unique ID generated for this device by Cognito.

$sel:httpStatus:RegisterDeviceResponse', registerDeviceResponse_httpStatus - The response's http status code.

BulkPublish

data BulkPublish

The input for the BulkPublish operation.

See: newBulkPublish smart constructor.

Constructors

BulkPublish' Text

newBulkPublish

Arguments

:: Text
$sel:identityPoolId:BulkPublish'

-> BulkPublish

Create a value of BulkPublish with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identityPoolId:BulkPublish', bulkPublish_identityPoolId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

data BulkPublishResponse

The output for the BulkPublish operation.

See: newBulkPublishResponse smart constructor.

Constructors

BulkPublishResponse' (Maybe Text) Int

newBulkPublishResponse

Arguments

:: Int
$sel:httpStatus:BulkPublishResponse'

-> BulkPublishResponse

Create a value of BulkPublishResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identityPoolId:BulkPublish', bulkPublishResponse_identityPoolId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

$sel:httpStatus:BulkPublishResponse', bulkPublishResponse_httpStatus - The response's http status code.

SubscribeToDataset

data SubscribeToDataset

A request to SubscribeToDatasetRequest.

See: newSubscribeToDataset smart constructor.

Constructors

SubscribeToDataset' Text Text Text Text

newSubscribeToDataset

Arguments

:: Text
$sel:identityPoolId:SubscribeToDataset'

-> Text
$sel:identityId:SubscribeToDataset'

-> Text
$sel:datasetName:SubscribeToDataset'

-> Text
$sel:deviceId:SubscribeToDataset'

-> SubscribeToDataset

Create a value of SubscribeToDataset with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identityPoolId:SubscribeToDataset', subscribeToDataset_identityPoolId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. The ID of the pool to which the identity belongs.

$sel:identityId:SubscribeToDataset', subscribeToDataset_identityId - Unique ID for this identity.

$sel:datasetName:SubscribeToDataset', subscribeToDataset_datasetName - The name of the dataset to subcribe to.

$sel:deviceId:SubscribeToDataset', subscribeToDataset_deviceId - The unique ID generated for this device by Cognito.

data SubscribeToDatasetResponse

Response to a SubscribeToDataset request.

See: newSubscribeToDatasetResponse smart constructor.

Constructors

SubscribeToDatasetResponse' Int

newSubscribeToDatasetResponse

Arguments

:: Int
$sel:httpStatus:SubscribeToDatasetResponse'

-> SubscribeToDatasetResponse

Create a value of SubscribeToDatasetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:SubscribeToDatasetResponse', subscribeToDatasetResponse_httpStatus - The response's http status code.

DescribeIdentityUsage

data DescribeIdentityUsage

A request for information about the usage of an identity pool.

See: newDescribeIdentityUsage smart constructor.

Constructors

DescribeIdentityUsage' Text Text

newDescribeIdentityUsage

Arguments

:: Text
$sel:identityPoolId:DescribeIdentityUsage'

-> Text
$sel:identityId:DescribeIdentityUsage'

-> DescribeIdentityUsage

Create a value of DescribeIdentityUsage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identityPoolId:DescribeIdentityUsage', describeIdentityUsage_identityPoolId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

$sel:identityId:DescribeIdentityUsage', describeIdentityUsage_identityId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

data DescribeIdentityUsageResponse

The response to a successful DescribeIdentityUsage request.

See: newDescribeIdentityUsageResponse smart constructor.

Constructors

DescribeIdentityUsageResponse' (Maybe IdentityUsage) Int

newDescribeIdentityUsageResponse

Arguments

:: Int
$sel:httpStatus:DescribeIdentityUsageResponse'

-> DescribeIdentityUsageResponse

Create a value of DescribeIdentityUsageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identityUsage:DescribeIdentityUsageResponse', describeIdentityUsageResponse_identityUsage - Usage information for the identity.

$sel:httpStatus:DescribeIdentityUsageResponse', describeIdentityUsageResponse_httpStatus - The response's http status code.

SetIdentityPoolConfiguration

data SetIdentityPoolConfiguration

The input for the SetIdentityPoolConfiguration operation.

See: newSetIdentityPoolConfiguration smart constructor.

Constructors

SetIdentityPoolConfiguration' (Maybe PushSync) (Maybe CognitoStreams) Text

newSetIdentityPoolConfiguration

Arguments

:: Text
$sel:identityPoolId:SetIdentityPoolConfiguration'

-> SetIdentityPoolConfiguration

Create a value of SetIdentityPoolConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pushSync:SetIdentityPoolConfiguration', setIdentityPoolConfiguration_pushSync - Options to apply to this identity pool for push synchronization.

$sel:cognitoStreams:SetIdentityPoolConfiguration', setIdentityPoolConfiguration_cognitoStreams - Options to apply to this identity pool for Amazon Cognito streams.

$sel:identityPoolId:SetIdentityPoolConfiguration', setIdentityPoolConfiguration_identityPoolId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. This is the ID of the pool to modify.

data SetIdentityPoolConfigurationResponse

The output for the SetIdentityPoolConfiguration operation

See: newSetIdentityPoolConfigurationResponse smart constructor.

Constructors

SetIdentityPoolConfigurationResponse' (Maybe Text) (Maybe PushSync) (Maybe CognitoStreams) Int

newSetIdentityPoolConfigurationResponse

Arguments

:: Int
$sel:httpStatus:SetIdentityPoolConfigurationResponse'

-> SetIdentityPoolConfigurationResponse

Create a value of SetIdentityPoolConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identityPoolId:SetIdentityPoolConfiguration', setIdentityPoolConfigurationResponse_identityPoolId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito.

$sel:pushSync:SetIdentityPoolConfiguration', setIdentityPoolConfigurationResponse_pushSync - Options to apply to this identity pool for push synchronization.

$sel:cognitoStreams:SetIdentityPoolConfiguration', setIdentityPoolConfigurationResponse_cognitoStreams - Options to apply to this identity pool for Amazon Cognito streams.

$sel:httpStatus:SetIdentityPoolConfigurationResponse', setIdentityPoolConfigurationResponse_httpStatus - The response's http status code.

GetBulkPublishDetails

data GetBulkPublishDetails

The input for the GetBulkPublishDetails operation.

See: newGetBulkPublishDetails smart constructor.

Constructors

GetBulkPublishDetails' Text

newGetBulkPublishDetails

Arguments

:: Text
$sel:identityPoolId:GetBulkPublishDetails'

-> GetBulkPublishDetails

Create a value of GetBulkPublishDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identityPoolId:GetBulkPublishDetails', getBulkPublishDetails_identityPoolId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

data GetBulkPublishDetailsResponse

The output for the GetBulkPublishDetails operation.

See: newGetBulkPublishDetailsResponse smart constructor.

Constructors

GetBulkPublishDetailsResponse' (Maybe Text) (Maybe POSIX) (Maybe Text) (Maybe POSIX) (Maybe BulkPublishStatus) Int

newGetBulkPublishDetailsResponse

Arguments

:: Int
$sel:httpStatus:GetBulkPublishDetailsResponse'

-> GetBulkPublishDetailsResponse

Create a value of GetBulkPublishDetailsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identityPoolId:GetBulkPublishDetails', getBulkPublishDetailsResponse_identityPoolId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

$sel:bulkPublishStartTime:GetBulkPublishDetailsResponse', getBulkPublishDetailsResponse_bulkPublishStartTime - The date/time at which the last bulk publish was initiated.

$sel:failureMessage:GetBulkPublishDetailsResponse', getBulkPublishDetailsResponse_failureMessage - If BulkPublishStatus is FAILED this field will contain the error message that caused the bulk publish to fail.

$sel:bulkPublishCompleteTime:GetBulkPublishDetailsResponse', getBulkPublishDetailsResponse_bulkPublishCompleteTime - If BulkPublishStatus is SUCCEEDED, the time the last bulk publish operation completed.

$sel:bulkPublishStatus:GetBulkPublishDetailsResponse', getBulkPublishDetailsResponse_bulkPublishStatus - Status of the last bulk publish operation, valid values are:

NOT_STARTED - No bulk publish has been requested for this identity pool

IN_PROGRESS - Data is being published to the configured stream

SUCCEEDED - All data for the identity pool has been published to the configured stream

FAILED - Some portion of the data has failed to publish, check FailureMessage for the cause.

$sel:httpStatus:GetBulkPublishDetailsResponse', getBulkPublishDetailsResponse_httpStatus - The response's http status code.

ListIdentityPoolUsage

data ListIdentityPoolUsage

A request for usage information on an identity pool.

See: newListIdentityPoolUsage smart constructor.

Constructors

ListIdentityPoolUsage' (Maybe Text) (Maybe Int)

newListIdentityPoolUsage :: ListIdentityPoolUsage

Create a value of ListIdentityPoolUsage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListIdentityPoolUsage', listIdentityPoolUsage_nextToken - A pagination token for obtaining the next page of results.

$sel:maxResults:ListIdentityPoolUsage', listIdentityPoolUsage_maxResults - The maximum number of results to be returned.

data ListIdentityPoolUsageResponse

Returned for a successful ListIdentityPoolUsage request.

See: newListIdentityPoolUsageResponse smart constructor.

Constructors

ListIdentityPoolUsageResponse' (Maybe [IdentityPoolUsage]) (Maybe Text) (Maybe Int) (Maybe Int) Int

newListIdentityPoolUsageResponse

Arguments

:: Int
$sel:httpStatus:ListIdentityPoolUsageResponse'

-> ListIdentityPoolUsageResponse

Create a value of ListIdentityPoolUsageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identityPoolUsages:ListIdentityPoolUsageResponse', listIdentityPoolUsageResponse_identityPoolUsages - Usage information for the identity pools.

$sel:nextToken:ListIdentityPoolUsage', listIdentityPoolUsageResponse_nextToken - A pagination token for obtaining the next page of results.

$sel:maxResults:ListIdentityPoolUsage', listIdentityPoolUsageResponse_maxResults - The maximum number of results to be returned.

$sel:count:ListIdentityPoolUsageResponse', listIdentityPoolUsageResponse_count - Total number of identities for the identity pool.

$sel:httpStatus:ListIdentityPoolUsageResponse', listIdentityPoolUsageResponse_httpStatus - The response's http status code.

SetCognitoEvents

data SetCognitoEvents

A request to configure Cognito Events"

"

See: newSetCognitoEvents smart constructor.

Constructors

SetCognitoEvents' Text (HashMap Text Text)

newSetCognitoEvents

Arguments

:: Text
$sel:identityPoolId:SetCognitoEvents'

-> SetCognitoEvents

Create a value of SetCognitoEvents with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identityPoolId:SetCognitoEvents', setCognitoEvents_identityPoolId - The Cognito Identity Pool to use when configuring Cognito Events

$sel:events:SetCognitoEvents', setCognitoEvents_events - The events to configure

data SetCognitoEventsResponse

See: newSetCognitoEventsResponse smart constructor.

Constructors

SetCognitoEventsResponse'

newSetCognitoEventsResponse :: SetCognitoEventsResponse

Create a value of SetCognitoEventsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

DescribeDataset

data DescribeDataset

A request for meta data about a dataset (creation date, number of records, size) by owner and dataset name.

See: newDescribeDataset smart constructor.

Constructors

DescribeDataset' Text Text Text

newDescribeDataset

Arguments

:: Text
$sel:identityPoolId:DescribeDataset'

-> Text
$sel:identityId:DescribeDataset'

-> Text
$sel:datasetName:DescribeDataset'

-> DescribeDataset

Create a value of DescribeDataset with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identityPoolId:DescribeDataset', describeDataset_identityPoolId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

$sel:identityId:DescribeDataset', describeDataset_identityId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

$sel:datasetName:DescribeDataset', describeDataset_datasetName - A string of up to 128 characters. Allowed characters are a-z, A-Z, 0-9, '_' (underscore), '-' (dash), and '.' (dot).

data DescribeDatasetResponse

Response to a successful DescribeDataset request.

See: newDescribeDatasetResponse smart constructor.

Constructors

DescribeDatasetResponse' (Maybe Dataset) Int

newDescribeDatasetResponse

Arguments

:: Int
$sel:httpStatus:DescribeDatasetResponse'

-> DescribeDatasetResponse

Create a value of DescribeDatasetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataset:DescribeDatasetResponse', describeDatasetResponse_dataset - Meta data for a collection of data for an identity. An identity can have multiple datasets. A dataset can be general or associated with a particular entity in an application (like a saved game). Datasets are automatically created if they don't exist. Data is synced by dataset, and a dataset can hold up to 1MB of key-value pairs.

$sel:httpStatus:DescribeDatasetResponse', describeDatasetResponse_httpStatus - The response's http status code.

UnsubscribeFromDataset

data UnsubscribeFromDataset

A request to UnsubscribeFromDataset.

See: newUnsubscribeFromDataset smart constructor.

Constructors

UnsubscribeFromDataset' Text Text Text Text

newUnsubscribeFromDataset

Arguments

:: Text
$sel:identityPoolId:UnsubscribeFromDataset'

-> Text
$sel:identityId:UnsubscribeFromDataset'

-> Text
$sel:datasetName:UnsubscribeFromDataset'

-> Text
$sel:deviceId:UnsubscribeFromDataset'

-> UnsubscribeFromDataset

Create a value of UnsubscribeFromDataset with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identityPoolId:UnsubscribeFromDataset', unsubscribeFromDataset_identityPoolId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. The ID of the pool to which this identity belongs.

$sel:identityId:UnsubscribeFromDataset', unsubscribeFromDataset_identityId - Unique ID for this identity.

$sel:datasetName:UnsubscribeFromDataset', unsubscribeFromDataset_datasetName - The name of the dataset from which to unsubcribe.

$sel:deviceId:UnsubscribeFromDataset', unsubscribeFromDataset_deviceId - The unique ID generated for this device by Cognito.

data UnsubscribeFromDatasetResponse

Response to an UnsubscribeFromDataset request.

See: newUnsubscribeFromDatasetResponse smart constructor.

Constructors

UnsubscribeFromDatasetResponse' Int

newUnsubscribeFromDatasetResponse

Arguments

:: Int
$sel:httpStatus:UnsubscribeFromDatasetResponse'

-> UnsubscribeFromDatasetResponse

Create a value of UnsubscribeFromDatasetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UnsubscribeFromDatasetResponse', unsubscribeFromDatasetResponse_httpStatus - The response's http status code.

GetIdentityPoolConfiguration

data GetIdentityPoolConfiguration

The input for the GetIdentityPoolConfiguration operation.

See: newGetIdentityPoolConfiguration smart constructor.

Constructors

GetIdentityPoolConfiguration' Text

newGetIdentityPoolConfiguration

Arguments

:: Text
$sel:identityPoolId:GetIdentityPoolConfiguration'

-> GetIdentityPoolConfiguration

Create a value of GetIdentityPoolConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identityPoolId:GetIdentityPoolConfiguration', getIdentityPoolConfiguration_identityPoolId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. This is the ID of the pool for which to return a configuration.

data GetIdentityPoolConfigurationResponse

The output for the GetIdentityPoolConfiguration operation.

See: newGetIdentityPoolConfigurationResponse smart constructor.

Constructors

GetIdentityPoolConfigurationResponse' (Maybe Text) (Maybe PushSync) (Maybe CognitoStreams) Int

newGetIdentityPoolConfigurationResponse

Arguments

:: Int
$sel:httpStatus:GetIdentityPoolConfigurationResponse'

-> GetIdentityPoolConfigurationResponse

Create a value of GetIdentityPoolConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identityPoolId:GetIdentityPoolConfiguration', getIdentityPoolConfigurationResponse_identityPoolId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito.

$sel:pushSync:GetIdentityPoolConfigurationResponse', getIdentityPoolConfigurationResponse_pushSync - Options to apply to this identity pool for push synchronization.

$sel:cognitoStreams:GetIdentityPoolConfigurationResponse', getIdentityPoolConfigurationResponse_cognitoStreams - Options to apply to this identity pool for Amazon Cognito streams.

$sel:httpStatus:GetIdentityPoolConfigurationResponse', getIdentityPoolConfigurationResponse_httpStatus - The response's http status code.

GetCognitoEvents

data GetCognitoEvents

A request for a list of the configured Cognito Events

See: newGetCognitoEvents smart constructor.

Constructors

GetCognitoEvents' Text

newGetCognitoEvents

Arguments

:: Text
$sel:identityPoolId:GetCognitoEvents'

-> GetCognitoEvents

Create a value of GetCognitoEvents with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identityPoolId:GetCognitoEvents', getCognitoEvents_identityPoolId - The Cognito Identity Pool ID for the request

data GetCognitoEventsResponse

The response from the GetCognitoEvents request

See: newGetCognitoEventsResponse smart constructor.

Constructors

GetCognitoEventsResponse' (Maybe (HashMap Text Text)) Int

newGetCognitoEventsResponse

Arguments

:: Int
$sel:httpStatus:GetCognitoEventsResponse'

-> GetCognitoEventsResponse

Create a value of GetCognitoEventsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:events:GetCognitoEventsResponse', getCognitoEventsResponse_events - The Cognito Events returned from the GetCognitoEvents request

$sel:httpStatus:GetCognitoEventsResponse', getCognitoEventsResponse_httpStatus - The response's http status code.

ListDatasets

data ListDatasets

Request for a list of datasets for an identity.

See: newListDatasets smart constructor.

Constructors

ListDatasets' (Maybe Text) (Maybe Int) Text Text

newListDatasets

Arguments

:: Text
$sel:identityId:ListDatasets'

-> Text
$sel:identityPoolId:ListDatasets'

-> ListDatasets

Create a value of ListDatasets with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDatasets', listDatasets_nextToken - A pagination token for obtaining the next page of results.

$sel:maxResults:ListDatasets', listDatasets_maxResults - The maximum number of results to be returned.

$sel:identityId:ListDatasets', listDatasets_identityId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

$sel:identityPoolId:ListDatasets', listDatasets_identityPoolId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

data ListDatasetsResponse

Returned for a successful ListDatasets request.

See: newListDatasetsResponse smart constructor.

Constructors

ListDatasetsResponse' (Maybe Text) (Maybe [Dataset]) (Maybe Int) Int

newListDatasetsResponse

Arguments

:: Int
$sel:httpStatus:ListDatasetsResponse'

-> ListDatasetsResponse

Create a value of ListDatasetsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDatasets', listDatasetsResponse_nextToken - A pagination token for obtaining the next page of results.

$sel:datasets:ListDatasetsResponse', listDatasetsResponse_datasets - A set of datasets.

$sel:count:ListDatasetsResponse', listDatasetsResponse_count - Number of datasets returned.

$sel:httpStatus:ListDatasetsResponse', listDatasetsResponse_httpStatus - The response's http status code.

DeleteDataset

data DeleteDataset

A request to delete the specific dataset.

See: newDeleteDataset smart constructor.

Constructors

DeleteDataset' Text Text Text

newDeleteDataset

Arguments

:: Text
$sel:identityPoolId:DeleteDataset'

-> Text
$sel:identityId:DeleteDataset'

-> Text
$sel:datasetName:DeleteDataset'

-> DeleteDataset

Create a value of DeleteDataset with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identityPoolId:DeleteDataset', deleteDataset_identityPoolId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

$sel:identityId:DeleteDataset', deleteDataset_identityId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

$sel:datasetName:DeleteDataset', deleteDataset_datasetName - A string of up to 128 characters. Allowed characters are a-z, A-Z, 0-9, '_' (underscore), '-' (dash), and '.' (dot).

data DeleteDatasetResponse

Response to a successful DeleteDataset request.

See: newDeleteDatasetResponse smart constructor.

Constructors

DeleteDatasetResponse' (Maybe Dataset) Int

newDeleteDatasetResponse

Arguments

:: Int
$sel:httpStatus:DeleteDatasetResponse'

-> DeleteDatasetResponse

Create a value of DeleteDatasetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataset:DeleteDatasetResponse', deleteDatasetResponse_dataset - A collection of data for an identity pool. An identity pool can have multiple datasets. A dataset is per identity and can be general or associated with a particular entity in an application (like a saved game). Datasets are automatically created if they don't exist. Data is synced by dataset, and a dataset can hold up to 1MB of key-value pairs.

$sel:httpStatus:DeleteDatasetResponse', deleteDatasetResponse_httpStatus - The response's http status code.

DescribeIdentityPoolUsage

data DescribeIdentityPoolUsage

A request for usage information about the identity pool.

See: newDescribeIdentityPoolUsage smart constructor.

Constructors

DescribeIdentityPoolUsage' Text

newDescribeIdentityPoolUsage

Arguments

:: Text
$sel:identityPoolId:DescribeIdentityPoolUsage'

-> DescribeIdentityPoolUsage

Create a value of DescribeIdentityPoolUsage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identityPoolId:DescribeIdentityPoolUsage', describeIdentityPoolUsage_identityPoolId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

data DescribeIdentityPoolUsageResponse

Response to a successful DescribeIdentityPoolUsage request.

See: newDescribeIdentityPoolUsageResponse smart constructor.

Constructors

DescribeIdentityPoolUsageResponse' (Maybe IdentityPoolUsage) Int

newDescribeIdentityPoolUsageResponse

Arguments

:: Int
$sel:httpStatus:DescribeIdentityPoolUsageResponse'

-> DescribeIdentityPoolUsageResponse

Create a value of DescribeIdentityPoolUsageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identityPoolUsage:DescribeIdentityPoolUsageResponse', describeIdentityPoolUsageResponse_identityPoolUsage - Information about the usage of the identity pool.

$sel:httpStatus:DescribeIdentityPoolUsageResponse', describeIdentityPoolUsageResponse_httpStatus - The response's http status code.

Types

BulkPublishStatus

newtype BulkPublishStatus

Constructors

BulkPublishStatus'

Fields

fromBulkPublishStatus :: Text

Bundled Patterns

pattern BulkPublishStatus_FAILED :: BulkPublishStatus

pattern BulkPublishStatus_IN_PROGRESS :: BulkPublishStatus

pattern BulkPublishStatus_NOT_STARTED :: BulkPublishStatus

pattern BulkPublishStatus_SUCCEEDED :: BulkPublishStatus

Operation

newtype Operation

Constructors

Operation'

Fields

fromOperation :: Text

Bundled Patterns

pattern Operation_Remove :: Operation

pattern Operation_Replace :: Operation

Platform

newtype Platform

Constructors

Platform'

Fields

fromPlatform :: Text

Bundled Patterns

pattern Platform_ADM :: Platform

pattern Platform_APNS :: Platform

pattern Platform_APNS_SANDBOX :: Platform

pattern Platform_GCM :: Platform

StreamingStatus

newtype StreamingStatus

Constructors

StreamingStatus'

Fields

fromStreamingStatus :: Text

Bundled Patterns

pattern StreamingStatus_DISABLED :: StreamingStatus

pattern StreamingStatus_ENABLED :: StreamingStatus

CognitoStreams

data CognitoStreams

Configuration options for configure Cognito streams.

See: newCognitoStreams smart constructor.

Constructors

CognitoStreams' (Maybe Text) (Maybe Text) (Maybe StreamingStatus)

newCognitoStreams :: CognitoStreams

Create a value of CognitoStreams with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:roleArn:CognitoStreams', cognitoStreams_roleArn - The ARN of the role Amazon Cognito can assume in order to publish to the stream. This role must grant access to Amazon Cognito (cognito-sync) to invoke PutRecord on your Cognito stream.

$sel:streamName:CognitoStreams', cognitoStreams_streamName - The name of the Cognito stream to receive updates. This stream must be in the developers account and in the same region as the identity pool.

$sel:streamingStatus:CognitoStreams', cognitoStreams_streamingStatus - Status of the Cognito streams. Valid values are:

ENABLED - Streaming of updates to identity pool is enabled.

DISABLED - Streaming of updates to identity pool is disabled. Bulk publish will also fail if StreamingStatus is DISABLED.

Dataset

data Dataset

A collection of data for an identity pool. An identity pool can have multiple datasets. A dataset is per identity and can be general or associated with a particular entity in an application (like a saved game). Datasets are automatically created if they don't exist. Data is synced by dataset, and a dataset can hold up to 1MB of key-value pairs.

See: newDataset smart constructor.

Constructors

Dataset' (Maybe POSIX) (Maybe Integer) (Maybe POSIX) (Maybe Integer) (Maybe Text) (Maybe Text) (Maybe Text)

newDataset :: Dataset

Create a value of Dataset with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedDate:Dataset', dataset_lastModifiedDate - Date when the dataset was last modified.

$sel:numRecords:Dataset', dataset_numRecords - Number of records in this dataset.

$sel:creationDate:Dataset', dataset_creationDate - Date on which the dataset was created.

$sel:dataStorage:Dataset', dataset_dataStorage - Total size in bytes of the records in this dataset.

$sel:identityId:Dataset', dataset_identityId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

$sel:lastModifiedBy:Dataset', dataset_lastModifiedBy - The device that made the last change to this dataset.

$sel:datasetName:Dataset', dataset_datasetName - A string of up to 128 characters. Allowed characters are a-z, A-Z, 0-9, '_' (underscore), '-' (dash), and '.' (dot).

IdentityPoolUsage

data IdentityPoolUsage

Usage information for the identity pool.

See: newIdentityPoolUsage smart constructor.

Constructors

IdentityPoolUsage' (Maybe POSIX) (Maybe Text) (Maybe Integer) (Maybe Integer)

newIdentityPoolUsage :: IdentityPoolUsage

Create a value of IdentityPoolUsage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedDate:IdentityPoolUsage', identityPoolUsage_lastModifiedDate - Date on which the identity pool was last modified.

$sel:identityPoolId:IdentityPoolUsage', identityPoolUsage_identityPoolId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

$sel:syncSessionsCount:IdentityPoolUsage', identityPoolUsage_syncSessionsCount - Number of sync sessions for the identity pool.

$sel:dataStorage:IdentityPoolUsage', identityPoolUsage_dataStorage - Data storage information for the identity pool.

IdentityUsage

data IdentityUsage

Usage information for the identity.

See: newIdentityUsage smart constructor.

Constructors

IdentityUsage' (Maybe POSIX) (Maybe Text) (Maybe Int) (Maybe Integer) (Maybe Text)

newIdentityUsage :: IdentityUsage

Create a value of IdentityUsage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedDate:IdentityUsage', identityUsage_lastModifiedDate - Date on which the identity was last modified.

$sel:identityPoolId:IdentityUsage', identityUsage_identityPoolId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

$sel:datasetCount:IdentityUsage', identityUsage_datasetCount - Number of datasets for the identity.

$sel:dataStorage:IdentityUsage', identityUsage_dataStorage - Total data storage for this identity.

$sel:identityId:IdentityUsage', identityUsage_identityId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

PushSync

data PushSync

Configuration options to be applied to the identity pool.

See: newPushSync smart constructor.

Constructors

PushSync' (Maybe Text) (Maybe [Text])

newPushSync :: PushSync

Create a value of PushSync with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:roleArn:PushSync', pushSync_roleArn - A role configured to allow Cognito to call SNS on behalf of the developer.

$sel:applicationArns:PushSync', pushSync_applicationArns - List of SNS platform application ARNs that could be used by clients.

Record

data Record

The basic data structure of a dataset.

See: newRecord smart constructor.

Constructors

Record' (Maybe Text) (Maybe POSIX) (Maybe POSIX) (Maybe Integer) (Maybe Text) (Maybe Text)

newRecord :: Record

Create a value of Record with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:Record', record_key - The key for the record.

$sel:deviceLastModifiedDate:Record', record_deviceLastModifiedDate - The last modified date of the client device.

$sel:lastModifiedDate:Record', record_lastModifiedDate - The date on which the record was last modified.

$sel:syncCount:Record', record_syncCount - The server sync count for this record.

$sel:value:Record', record_value - The value for the record.

$sel:lastModifiedBy:Record', record_lastModifiedBy - The user/device that made the last change to this record.

RecordPatch

data RecordPatch

An update operation for a record.

See: newRecordPatch smart constructor.

Constructors

RecordPatch' (Maybe POSIX) (Maybe Text) Operation Text Integer

newRecordPatch

Arguments

:: Operation
$sel:op:RecordPatch'

-> Text
$sel:key:RecordPatch'

-> Integer
$sel:syncCount:RecordPatch'

-> RecordPatch

Create a value of RecordPatch with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceLastModifiedDate:RecordPatch', recordPatch_deviceLastModifiedDate - The last modified date of the client device.

$sel:value:RecordPatch', recordPatch_value - The value associated with the record patch.

$sel:op:RecordPatch', recordPatch_op - An operation, either replace or remove.

$sel:key:RecordPatch', recordPatch_key - The key associated with the record patch.

$sel:syncCount:RecordPatch', recordPatch_syncCount - Last known server sync count for this record. Set to 0 if unknown.