Network.AWS.CognitoSync.DescribeDataset

Description

Gets meta data about a dataset by identity and dataset name. With Amazon Cognito Sync, each identity has access only to its own data. Thus, the credentials used to make this API call need to have access to the identity data.

This API can be called with temporary user credentials provided by Cognito Identity or with developer credentials. You should use Cognito Identity credentials to make this API call.

Creating a Request

data DescribeDataset

A request for meta data about a dataset (creation date, number of records, size) by owner and dataset name.

See: newDescribeDataset smart constructor.

Constructors

DescribeDataset'

Fields

identityPoolId :: Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

identityId :: Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

datasetName :: Text

A string of up to 128 characters. Allowed characters are a-z, A-Z, 0-9, '_' (underscore), '-' (dash), and '.' (dot).

newDescribeDataset

Arguments

:: Text
$sel:identityPoolId:DescribeDataset'

-> Text
$sel:identityId:DescribeDataset'

-> Text
$sel:datasetName:DescribeDataset'

-> DescribeDataset

Create a value of DescribeDataset with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identityPoolId:DescribeDataset', describeDataset_identityPoolId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

$sel:identityId:DescribeDataset', describeDataset_identityId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

$sel:datasetName:DescribeDataset', describeDataset_datasetName - A string of up to 128 characters. Allowed characters are a-z, A-Z, 0-9, '_' (underscore), '-' (dash), and '.' (dot).

Request Lenses

describeDataset_identityPoolId :: Lens' DescribeDataset Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

describeDataset_identityId :: Lens' DescribeDataset Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

describeDataset_datasetName :: Lens' DescribeDataset Text

A string of up to 128 characters. Allowed characters are a-z, A-Z, 0-9, '_' (underscore), '-' (dash), and '.' (dot).

Destructuring the Response

data DescribeDatasetResponse

Response to a successful DescribeDataset request.

See: newDescribeDatasetResponse smart constructor.

Constructors

DescribeDatasetResponse'

Fields

dataset :: Maybe Dataset

Meta data for a collection of data for an identity. An identity can have multiple datasets. A dataset can be general or associated with a particular entity in an application (like a saved game). Datasets are automatically created if they don't exist. Data is synced by dataset, and a dataset can hold up to 1MB of key-value pairs.

httpStatus :: Int

The response's http status code.

newDescribeDatasetResponse

Arguments

:: Int
$sel:httpStatus:DescribeDatasetResponse'

-> DescribeDatasetResponse

Create a value of DescribeDatasetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataset:DescribeDatasetResponse', describeDatasetResponse_dataset - Meta data for a collection of data for an identity. An identity can have multiple datasets. A dataset can be general or associated with a particular entity in an application (like a saved game). Datasets are automatically created if they don't exist. Data is synced by dataset, and a dataset can hold up to 1MB of key-value pairs.

$sel:httpStatus:DescribeDatasetResponse', describeDatasetResponse_httpStatus - The response's http status code.

Response Lenses

describeDatasetResponse_dataset :: Lens' DescribeDatasetResponse (Maybe Dataset)

Meta data for a collection of data for an identity. An identity can have multiple datasets. A dataset can be general or associated with a particular entity in an application (like a saved game). Datasets are automatically created if they don't exist. Data is synced by dataset, and a dataset can hold up to 1MB of key-value pairs.

describeDatasetResponse_httpStatus :: Lens' DescribeDatasetResponse Int

The response's http status code.