Network.AWS.CognitoSync.GetBulkPublishDetails

Description

Get the status of the last BulkPublish operation for an identity pool.

This API can only be called with developer credentials. You cannot call this API with the temporary user credentials provided by Cognito Identity.

Creating a Request

data GetBulkPublishDetails

The input for the GetBulkPublishDetails operation.

See: newGetBulkPublishDetails smart constructor.

Constructors

GetBulkPublishDetails'

Fields

identityPoolId :: Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

newGetBulkPublishDetails

Arguments

:: Text
$sel:identityPoolId:GetBulkPublishDetails'

-> GetBulkPublishDetails

Create a value of GetBulkPublishDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identityPoolId:GetBulkPublishDetails', getBulkPublishDetails_identityPoolId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

Request Lenses

getBulkPublishDetails_identityPoolId :: Lens' GetBulkPublishDetails Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

Destructuring the Response

data GetBulkPublishDetailsResponse

The output for the GetBulkPublishDetails operation.

See: newGetBulkPublishDetailsResponse smart constructor.

Constructors

GetBulkPublishDetailsResponse'

Fields

identityPoolId :: Maybe Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

bulkPublishStartTime :: Maybe POSIX

The date/time at which the last bulk publish was initiated.

failureMessage :: Maybe Text

If BulkPublishStatus is FAILED this field will contain the error message that caused the bulk publish to fail.

bulkPublishCompleteTime :: Maybe POSIX

If BulkPublishStatus is SUCCEEDED, the time the last bulk publish operation completed.

bulkPublishStatus :: Maybe BulkPublishStatus

Status of the last bulk publish operation, valid values are:

NOT_STARTED - No bulk publish has been requested for this identity pool

IN_PROGRESS - Data is being published to the configured stream

SUCCEEDED - All data for the identity pool has been published to the configured stream

FAILED - Some portion of the data has failed to publish, check FailureMessage for the cause.

httpStatus :: Int

The response's http status code.

newGetBulkPublishDetailsResponse

Arguments

:: Int
$sel:httpStatus:GetBulkPublishDetailsResponse'

-> GetBulkPublishDetailsResponse

Create a value of GetBulkPublishDetailsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identityPoolId:GetBulkPublishDetails', getBulkPublishDetailsResponse_identityPoolId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

$sel:bulkPublishStartTime:GetBulkPublishDetailsResponse', getBulkPublishDetailsResponse_bulkPublishStartTime - The date/time at which the last bulk publish was initiated.

$sel:failureMessage:GetBulkPublishDetailsResponse', getBulkPublishDetailsResponse_failureMessage - If BulkPublishStatus is FAILED this field will contain the error message that caused the bulk publish to fail.

$sel:bulkPublishCompleteTime:GetBulkPublishDetailsResponse', getBulkPublishDetailsResponse_bulkPublishCompleteTime - If BulkPublishStatus is SUCCEEDED, the time the last bulk publish operation completed.

$sel:bulkPublishStatus:GetBulkPublishDetailsResponse', getBulkPublishDetailsResponse_bulkPublishStatus - Status of the last bulk publish operation, valid values are:

NOT_STARTED - No bulk publish has been requested for this identity pool

IN_PROGRESS - Data is being published to the configured stream

SUCCEEDED - All data for the identity pool has been published to the configured stream

FAILED - Some portion of the data has failed to publish, check FailureMessage for the cause.

$sel:httpStatus:GetBulkPublishDetailsResponse', getBulkPublishDetailsResponse_httpStatus - The response's http status code.

Response Lenses

getBulkPublishDetailsResponse_identityPoolId :: Lens' GetBulkPublishDetailsResponse (Maybe Text)

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

getBulkPublishDetailsResponse_bulkPublishStartTime :: Lens' GetBulkPublishDetailsResponse (Maybe UTCTime)

The date/time at which the last bulk publish was initiated.

getBulkPublishDetailsResponse_failureMessage :: Lens' GetBulkPublishDetailsResponse (Maybe Text)

If BulkPublishStatus is FAILED this field will contain the error message that caused the bulk publish to fail.

getBulkPublishDetailsResponse_bulkPublishCompleteTime :: Lens' GetBulkPublishDetailsResponse (Maybe UTCTime)

If BulkPublishStatus is SUCCEEDED, the time the last bulk publish operation completed.

getBulkPublishDetailsResponse_bulkPublishStatus :: Lens' GetBulkPublishDetailsResponse (Maybe BulkPublishStatus)

Status of the last bulk publish operation, valid values are:

NOT_STARTED - No bulk publish has been requested for this identity pool

IN_PROGRESS - Data is being published to the configured stream

SUCCEEDED - All data for the identity pool has been published to the configured stream

FAILED - Some portion of the data has failed to publish, check FailureMessage for the cause.

getBulkPublishDetailsResponse_httpStatus :: Lens' GetBulkPublishDetailsResponse Int

The response's http status code.