Network.AWS.CognitoSync.GetCognitoEvents

Description

Gets the events and the corresponding Lambda functions associated with an identity pool.

This API can only be called with developer credentials. You cannot call this API with the temporary user credentials provided by Cognito Identity.

Creating a Request

data GetCognitoEvents

A request for a list of the configured Cognito Events

See: newGetCognitoEvents smart constructor.

Constructors

GetCognitoEvents'

Fields

identityPoolId :: Text

The Cognito Identity Pool ID for the request

newGetCognitoEvents

Arguments

:: Text
$sel:identityPoolId:GetCognitoEvents'

-> GetCognitoEvents

Create a value of GetCognitoEvents with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identityPoolId:GetCognitoEvents', getCognitoEvents_identityPoolId - The Cognito Identity Pool ID for the request

Request Lenses

getCognitoEvents_identityPoolId :: Lens' GetCognitoEvents Text

The Cognito Identity Pool ID for the request

Destructuring the Response

data GetCognitoEventsResponse

The response from the GetCognitoEvents request

See: newGetCognitoEventsResponse smart constructor.

Constructors

GetCognitoEventsResponse'

Fields

events :: Maybe (HashMap Text Text)

The Cognito Events returned from the GetCognitoEvents request

httpStatus :: Int

The response's http status code.

newGetCognitoEventsResponse

Arguments

:: Int
$sel:httpStatus:GetCognitoEventsResponse'

-> GetCognitoEventsResponse

Create a value of GetCognitoEventsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:events:GetCognitoEventsResponse', getCognitoEventsResponse_events - The Cognito Events returned from the GetCognitoEvents request

$sel:httpStatus:GetCognitoEventsResponse', getCognitoEventsResponse_httpStatus - The response's http status code.

Response Lenses

getCognitoEventsResponse_events :: Lens' GetCognitoEventsResponse (Maybe (HashMap Text Text))

The Cognito Events returned from the GetCognitoEvents request

getCognitoEventsResponse_httpStatus :: Lens' GetCognitoEventsResponse Int

The response's http status code.