Network.AWS.CognitoSync.Lens

Description

Operations

UpdateRecords

updateRecords_recordPatches :: Lens' UpdateRecords (Maybe [RecordPatch])

A list of patch operations.

updateRecords_deviceId :: Lens' UpdateRecords (Maybe Text)

The unique ID generated for this device by Cognito.

updateRecords_clientContext :: Lens' UpdateRecords (Maybe Text)

Intended to supply a device ID that will populate the lastModifiedBy field referenced in other methods. The ClientContext field is not yet implemented.

updateRecords_identityPoolId :: Lens' UpdateRecords Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

updateRecords_identityId :: Lens' UpdateRecords Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

updateRecords_datasetName :: Lens' UpdateRecords Text

A string of up to 128 characters. Allowed characters are a-z, A-Z, 0-9, '_' (underscore), '-' (dash), and '.' (dot).

updateRecords_syncSessionToken :: Lens' UpdateRecords Text

The SyncSessionToken returned by a previous call to ListRecords for this dataset and identity.

updateRecordsResponse_records :: Lens' UpdateRecordsResponse (Maybe [Record])

A list of records that have been updated.

updateRecordsResponse_httpStatus :: Lens' UpdateRecordsResponse Int

The response's http status code.

ListRecords

listRecords_nextToken :: Lens' ListRecords (Maybe Text)

A pagination token for obtaining the next page of results.

listRecords_maxResults :: Lens' ListRecords (Maybe Int)

The maximum number of results to be returned.

listRecords_lastSyncCount :: Lens' ListRecords (Maybe Integer)

The last server sync count for this record.

listRecords_syncSessionToken :: Lens' ListRecords (Maybe Text)

A token containing a session ID, identity ID, and expiration.

listRecords_identityPoolId :: Lens' ListRecords Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

listRecords_identityId :: Lens' ListRecords Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

listRecords_datasetName :: Lens' ListRecords Text

A string of up to 128 characters. Allowed characters are a-z, A-Z, 0-9, '_' (underscore), '-' (dash), and '.' (dot).

listRecordsResponse_nextToken :: Lens' ListRecordsResponse (Maybe Text)

A pagination token for obtaining the next page of results.

listRecordsResponse_datasetSyncCount :: Lens' ListRecordsResponse (Maybe Integer)

Server sync count for this dataset.

listRecordsResponse_records :: Lens' ListRecordsResponse (Maybe [Record])

A list of all records.

listRecordsResponse_datasetDeletedAfterRequestedSyncCount :: Lens' ListRecordsResponse (Maybe Bool)

A boolean value specifying whether to delete the dataset locally.

listRecordsResponse_count :: Lens' ListRecordsResponse (Maybe Int)

Total number of records.

listRecordsResponse_lastModifiedBy :: Lens' ListRecordsResponse (Maybe Text)

The user/device that made the last change to this record.

listRecordsResponse_datasetExists :: Lens' ListRecordsResponse (Maybe Bool)

Indicates whether the dataset exists.

listRecordsResponse_syncSessionToken :: Lens' ListRecordsResponse (Maybe Text)

A token containing a session ID, identity ID, and expiration.

listRecordsResponse_mergedDatasetNames :: Lens' ListRecordsResponse (Maybe [Text])

Names of merged datasets.

listRecordsResponse_httpStatus :: Lens' ListRecordsResponse Int

The response's http status code.

RegisterDevice

registerDevice_identityPoolId :: Lens' RegisterDevice Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. Here, the ID of the pool that the identity belongs to.

registerDevice_identityId :: Lens' RegisterDevice Text

The unique ID for this identity.

registerDevice_platform :: Lens' RegisterDevice Platform

The SNS platform type (e.g. GCM, SDM, APNS, APNS_SANDBOX).

registerDevice_token :: Lens' RegisterDevice Text

The push token.

registerDeviceResponse_deviceId :: Lens' RegisterDeviceResponse (Maybe Text)

The unique ID generated for this device by Cognito.

registerDeviceResponse_httpStatus :: Lens' RegisterDeviceResponse Int

The response's http status code.

BulkPublish

bulkPublish_identityPoolId :: Lens' BulkPublish Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

bulkPublishResponse_identityPoolId :: Lens' BulkPublishResponse (Maybe Text)

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

bulkPublishResponse_httpStatus :: Lens' BulkPublishResponse Int

The response's http status code.

SubscribeToDataset

subscribeToDataset_identityPoolId :: Lens' SubscribeToDataset Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. The ID of the pool to which the identity belongs.

subscribeToDataset_identityId :: Lens' SubscribeToDataset Text

Unique ID for this identity.

subscribeToDataset_datasetName :: Lens' SubscribeToDataset Text

The name of the dataset to subcribe to.

subscribeToDataset_deviceId :: Lens' SubscribeToDataset Text

The unique ID generated for this device by Cognito.

subscribeToDatasetResponse_httpStatus :: Lens' SubscribeToDatasetResponse Int

The response's http status code.

DescribeIdentityUsage

describeIdentityUsage_identityPoolId :: Lens' DescribeIdentityUsage Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

describeIdentityUsage_identityId :: Lens' DescribeIdentityUsage Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

describeIdentityUsageResponse_identityUsage :: Lens' DescribeIdentityUsageResponse (Maybe IdentityUsage)

Usage information for the identity.

describeIdentityUsageResponse_httpStatus :: Lens' DescribeIdentityUsageResponse Int

The response's http status code.

SetIdentityPoolConfiguration

setIdentityPoolConfiguration_pushSync :: Lens' SetIdentityPoolConfiguration (Maybe PushSync)

Options to apply to this identity pool for push synchronization.

setIdentityPoolConfiguration_cognitoStreams :: Lens' SetIdentityPoolConfiguration (Maybe CognitoStreams)

Options to apply to this identity pool for Amazon Cognito streams.

setIdentityPoolConfiguration_identityPoolId :: Lens' SetIdentityPoolConfiguration Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. This is the ID of the pool to modify.

setIdentityPoolConfigurationResponse_identityPoolId :: Lens' SetIdentityPoolConfigurationResponse (Maybe Text)

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito.

setIdentityPoolConfigurationResponse_pushSync :: Lens' SetIdentityPoolConfigurationResponse (Maybe PushSync)

Options to apply to this identity pool for push synchronization.

setIdentityPoolConfigurationResponse_cognitoStreams :: Lens' SetIdentityPoolConfigurationResponse (Maybe CognitoStreams)

Options to apply to this identity pool for Amazon Cognito streams.

setIdentityPoolConfigurationResponse_httpStatus :: Lens' SetIdentityPoolConfigurationResponse Int

The response's http status code.

GetBulkPublishDetails

getBulkPublishDetails_identityPoolId :: Lens' GetBulkPublishDetails Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

getBulkPublishDetailsResponse_identityPoolId :: Lens' GetBulkPublishDetailsResponse (Maybe Text)

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

getBulkPublishDetailsResponse_bulkPublishStartTime :: Lens' GetBulkPublishDetailsResponse (Maybe UTCTime)

The date/time at which the last bulk publish was initiated.

getBulkPublishDetailsResponse_failureMessage :: Lens' GetBulkPublishDetailsResponse (Maybe Text)

If BulkPublishStatus is FAILED this field will contain the error message that caused the bulk publish to fail.

getBulkPublishDetailsResponse_bulkPublishCompleteTime :: Lens' GetBulkPublishDetailsResponse (Maybe UTCTime)

If BulkPublishStatus is SUCCEEDED, the time the last bulk publish operation completed.

getBulkPublishDetailsResponse_bulkPublishStatus :: Lens' GetBulkPublishDetailsResponse (Maybe BulkPublishStatus)

Status of the last bulk publish operation, valid values are:

NOT_STARTED - No bulk publish has been requested for this identity pool

IN_PROGRESS - Data is being published to the configured stream

SUCCEEDED - All data for the identity pool has been published to the configured stream

FAILED - Some portion of the data has failed to publish, check FailureMessage for the cause.

getBulkPublishDetailsResponse_httpStatus :: Lens' GetBulkPublishDetailsResponse Int

The response's http status code.

ListIdentityPoolUsage

listIdentityPoolUsage_nextToken :: Lens' ListIdentityPoolUsage (Maybe Text)

A pagination token for obtaining the next page of results.

listIdentityPoolUsage_maxResults :: Lens' ListIdentityPoolUsage (Maybe Int)

The maximum number of results to be returned.

listIdentityPoolUsageResponse_identityPoolUsages :: Lens' ListIdentityPoolUsageResponse (Maybe [IdentityPoolUsage])

Usage information for the identity pools.

listIdentityPoolUsageResponse_nextToken :: Lens' ListIdentityPoolUsageResponse (Maybe Text)

A pagination token for obtaining the next page of results.

listIdentityPoolUsageResponse_maxResults :: Lens' ListIdentityPoolUsageResponse (Maybe Int)

The maximum number of results to be returned.

listIdentityPoolUsageResponse_count :: Lens' ListIdentityPoolUsageResponse (Maybe Int)

Total number of identities for the identity pool.

listIdentityPoolUsageResponse_httpStatus :: Lens' ListIdentityPoolUsageResponse Int

The response's http status code.

SetCognitoEvents

setCognitoEvents_identityPoolId :: Lens' SetCognitoEvents Text

The Cognito Identity Pool to use when configuring Cognito Events

setCognitoEvents_events :: Lens' SetCognitoEvents (HashMap Text Text)

The events to configure

DescribeDataset

describeDataset_identityPoolId :: Lens' DescribeDataset Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

describeDataset_identityId :: Lens' DescribeDataset Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

describeDataset_datasetName :: Lens' DescribeDataset Text

A string of up to 128 characters. Allowed characters are a-z, A-Z, 0-9, '_' (underscore), '-' (dash), and '.' (dot).

describeDatasetResponse_dataset :: Lens' DescribeDatasetResponse (Maybe Dataset)

Meta data for a collection of data for an identity. An identity can have multiple datasets. A dataset can be general or associated with a particular entity in an application (like a saved game). Datasets are automatically created if they don't exist. Data is synced by dataset, and a dataset can hold up to 1MB of key-value pairs.

describeDatasetResponse_httpStatus :: Lens' DescribeDatasetResponse Int

The response's http status code.

UnsubscribeFromDataset

unsubscribeFromDataset_identityPoolId :: Lens' UnsubscribeFromDataset Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. The ID of the pool to which this identity belongs.

unsubscribeFromDataset_identityId :: Lens' UnsubscribeFromDataset Text

Unique ID for this identity.

unsubscribeFromDataset_datasetName :: Lens' UnsubscribeFromDataset Text

The name of the dataset from which to unsubcribe.

unsubscribeFromDataset_deviceId :: Lens' UnsubscribeFromDataset Text

The unique ID generated for this device by Cognito.

unsubscribeFromDatasetResponse_httpStatus :: Lens' UnsubscribeFromDatasetResponse Int

The response's http status code.

GetIdentityPoolConfiguration

getIdentityPoolConfiguration_identityPoolId :: Lens' GetIdentityPoolConfiguration Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. This is the ID of the pool for which to return a configuration.

getIdentityPoolConfigurationResponse_identityPoolId :: Lens' GetIdentityPoolConfigurationResponse (Maybe Text)

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito.

getIdentityPoolConfigurationResponse_pushSync :: Lens' GetIdentityPoolConfigurationResponse (Maybe PushSync)

Options to apply to this identity pool for push synchronization.

getIdentityPoolConfigurationResponse_cognitoStreams :: Lens' GetIdentityPoolConfigurationResponse (Maybe CognitoStreams)

Options to apply to this identity pool for Amazon Cognito streams.

getIdentityPoolConfigurationResponse_httpStatus :: Lens' GetIdentityPoolConfigurationResponse Int

The response's http status code.

GetCognitoEvents

getCognitoEvents_identityPoolId :: Lens' GetCognitoEvents Text

The Cognito Identity Pool ID for the request

getCognitoEventsResponse_events :: Lens' GetCognitoEventsResponse (Maybe (HashMap Text Text))

The Cognito Events returned from the GetCognitoEvents request

getCognitoEventsResponse_httpStatus :: Lens' GetCognitoEventsResponse Int

The response's http status code.

ListDatasets

listDatasets_nextToken :: Lens' ListDatasets (Maybe Text)

A pagination token for obtaining the next page of results.

listDatasets_maxResults :: Lens' ListDatasets (Maybe Int)

The maximum number of results to be returned.

listDatasets_identityId :: Lens' ListDatasets Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

listDatasets_identityPoolId :: Lens' ListDatasets Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

listDatasetsResponse_nextToken :: Lens' ListDatasetsResponse (Maybe Text)

A pagination token for obtaining the next page of results.

listDatasetsResponse_datasets :: Lens' ListDatasetsResponse (Maybe [Dataset])

A set of datasets.

listDatasetsResponse_count :: Lens' ListDatasetsResponse (Maybe Int)

Number of datasets returned.

listDatasetsResponse_httpStatus :: Lens' ListDatasetsResponse Int

The response's http status code.

DeleteDataset

deleteDataset_identityPoolId :: Lens' DeleteDataset Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

deleteDataset_identityId :: Lens' DeleteDataset Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

deleteDataset_datasetName :: Lens' DeleteDataset Text

A string of up to 128 characters. Allowed characters are a-z, A-Z, 0-9, '_' (underscore), '-' (dash), and '.' (dot).

deleteDatasetResponse_dataset :: Lens' DeleteDatasetResponse (Maybe Dataset)

A collection of data for an identity pool. An identity pool can have multiple datasets. A dataset is per identity and can be general or associated with a particular entity in an application (like a saved game). Datasets are automatically created if they don't exist. Data is synced by dataset, and a dataset can hold up to 1MB of key-value pairs.

deleteDatasetResponse_httpStatus :: Lens' DeleteDatasetResponse Int

The response's http status code.

DescribeIdentityPoolUsage

describeIdentityPoolUsage_identityPoolId :: Lens' DescribeIdentityPoolUsage Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

describeIdentityPoolUsageResponse_identityPoolUsage :: Lens' DescribeIdentityPoolUsageResponse (Maybe IdentityPoolUsage)

Information about the usage of the identity pool.

describeIdentityPoolUsageResponse_httpStatus :: Lens' DescribeIdentityPoolUsageResponse Int

The response's http status code.

Types

CognitoStreams

cognitoStreams_roleArn :: Lens' CognitoStreams (Maybe Text)

The ARN of the role Amazon Cognito can assume in order to publish to the stream. This role must grant access to Amazon Cognito (cognito-sync) to invoke PutRecord on your Cognito stream.

cognitoStreams_streamName :: Lens' CognitoStreams (Maybe Text)

The name of the Cognito stream to receive updates. This stream must be in the developers account and in the same region as the identity pool.

cognitoStreams_streamingStatus :: Lens' CognitoStreams (Maybe StreamingStatus)

Status of the Cognito streams. Valid values are:

ENABLED - Streaming of updates to identity pool is enabled.

DISABLED - Streaming of updates to identity pool is disabled. Bulk publish will also fail if StreamingStatus is DISABLED.

Dataset

dataset_lastModifiedDate :: Lens' Dataset (Maybe UTCTime)

Date when the dataset was last modified.

dataset_numRecords :: Lens' Dataset (Maybe Integer)

Number of records in this dataset.

dataset_creationDate :: Lens' Dataset (Maybe UTCTime)

Date on which the dataset was created.

dataset_dataStorage :: Lens' Dataset (Maybe Integer)

Total size in bytes of the records in this dataset.

dataset_identityId :: Lens' Dataset (Maybe Text)

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

dataset_lastModifiedBy :: Lens' Dataset (Maybe Text)

The device that made the last change to this dataset.

dataset_datasetName :: Lens' Dataset (Maybe Text)

A string of up to 128 characters. Allowed characters are a-z, A-Z, 0-9, '_' (underscore), '-' (dash), and '.' (dot).

IdentityPoolUsage

identityPoolUsage_lastModifiedDate :: Lens' IdentityPoolUsage (Maybe UTCTime)

Date on which the identity pool was last modified.

identityPoolUsage_identityPoolId :: Lens' IdentityPoolUsage (Maybe Text)

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

identityPoolUsage_syncSessionsCount :: Lens' IdentityPoolUsage (Maybe Integer)

Number of sync sessions for the identity pool.

identityPoolUsage_dataStorage :: Lens' IdentityPoolUsage (Maybe Integer)

Data storage information for the identity pool.

IdentityUsage

identityUsage_lastModifiedDate :: Lens' IdentityUsage (Maybe UTCTime)

Date on which the identity was last modified.

identityUsage_identityPoolId :: Lens' IdentityUsage (Maybe Text)

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

identityUsage_datasetCount :: Lens' IdentityUsage (Maybe Int)

Number of datasets for the identity.

identityUsage_dataStorage :: Lens' IdentityUsage (Maybe Integer)

Total data storage for this identity.

identityUsage_identityId :: Lens' IdentityUsage (Maybe Text)

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

PushSync

pushSync_roleArn :: Lens' PushSync (Maybe Text)

A role configured to allow Cognito to call SNS on behalf of the developer.

pushSync_applicationArns :: Lens' PushSync (Maybe [Text])

List of SNS platform application ARNs that could be used by clients.

Record

record_key :: Lens' Record (Maybe Text)

The key for the record.

record_deviceLastModifiedDate :: Lens' Record (Maybe UTCTime)

The last modified date of the client device.

record_lastModifiedDate :: Lens' Record (Maybe UTCTime)

The date on which the record was last modified.

record_syncCount :: Lens' Record (Maybe Integer)

The server sync count for this record.

record_value :: Lens' Record (Maybe Text)

The value for the record.

record_lastModifiedBy :: Lens' Record (Maybe Text)

The user/device that made the last change to this record.

RecordPatch

recordPatch_deviceLastModifiedDate :: Lens' RecordPatch (Maybe UTCTime)

The last modified date of the client device.

recordPatch_value :: Lens' RecordPatch (Maybe Text)

The value associated with the record patch.

recordPatch_op :: Lens' RecordPatch Operation

An operation, either replace or remove.

recordPatch_key :: Lens' RecordPatch Text

The key associated with the record patch.

recordPatch_syncCount :: Lens' RecordPatch Integer

Last known server sync count for this record. Set to 0 if unknown.