Network.AWS.CognitoSync.ListDatasets

Description

Lists datasets for an identity. With Amazon Cognito Sync, each identity has access only to its own data. Thus, the credentials used to make this API call need to have access to the identity data.

ListDatasets can be called with temporary user credentials provided by Cognito Identity or with developer credentials. You should use the Cognito Identity credentials to make this API call.

Creating a Request

data ListDatasets

Request for a list of datasets for an identity.

See: newListDatasets smart constructor.

Constructors

ListDatasets'

Fields

nextToken :: Maybe Text

A pagination token for obtaining the next page of results.

maxResults :: Maybe Int

The maximum number of results to be returned.

identityId :: Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

identityPoolId :: Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

newListDatasets

Arguments

:: Text
$sel:identityId:ListDatasets'

-> Text
$sel:identityPoolId:ListDatasets'

-> ListDatasets

Create a value of ListDatasets with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDatasets', listDatasets_nextToken - A pagination token for obtaining the next page of results.

$sel:maxResults:ListDatasets', listDatasets_maxResults - The maximum number of results to be returned.

$sel:identityId:ListDatasets', listDatasets_identityId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

$sel:identityPoolId:ListDatasets', listDatasets_identityPoolId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

Request Lenses

listDatasets_nextToken :: Lens' ListDatasets (Maybe Text)

A pagination token for obtaining the next page of results.

listDatasets_maxResults :: Lens' ListDatasets (Maybe Int)

The maximum number of results to be returned.

listDatasets_identityId :: Lens' ListDatasets Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

listDatasets_identityPoolId :: Lens' ListDatasets Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

Destructuring the Response

data ListDatasetsResponse

Returned for a successful ListDatasets request.

See: newListDatasetsResponse smart constructor.

Constructors

ListDatasetsResponse'

Fields

nextToken :: Maybe Text

A pagination token for obtaining the next page of results.

datasets :: Maybe [Dataset]

A set of datasets.

count :: Maybe Int

Number of datasets returned.

httpStatus :: Int

The response's http status code.

newListDatasetsResponse

Arguments

:: Int
$sel:httpStatus:ListDatasetsResponse'

-> ListDatasetsResponse

Create a value of ListDatasetsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDatasets', listDatasetsResponse_nextToken - A pagination token for obtaining the next page of results.

$sel:datasets:ListDatasetsResponse', listDatasetsResponse_datasets - A set of datasets.

$sel:count:ListDatasetsResponse', listDatasetsResponse_count - Number of datasets returned.

$sel:httpStatus:ListDatasetsResponse', listDatasetsResponse_httpStatus - The response's http status code.

Response Lenses

listDatasetsResponse_nextToken :: Lens' ListDatasetsResponse (Maybe Text)

A pagination token for obtaining the next page of results.

listDatasetsResponse_datasets :: Lens' ListDatasetsResponse (Maybe [Dataset])

A set of datasets.

listDatasetsResponse_count :: Lens' ListDatasetsResponse (Maybe Int)

Number of datasets returned.

listDatasetsResponse_httpStatus :: Lens' ListDatasetsResponse Int

The response's http status code.