Network.AWS.CognitoSync.ListIdentityPoolUsage

Description

Gets a list of identity pools registered with Cognito.

ListIdentityPoolUsage can only be called with developer credentials. You cannot make this API call with the temporary user credentials provided by Cognito Identity.

Creating a Request

data ListIdentityPoolUsage

A request for usage information on an identity pool.

See: newListIdentityPoolUsage smart constructor.

Constructors

ListIdentityPoolUsage'

Fields

nextToken :: Maybe Text

A pagination token for obtaining the next page of results.

maxResults :: Maybe Int

The maximum number of results to be returned.

newListIdentityPoolUsage :: ListIdentityPoolUsage

Create a value of ListIdentityPoolUsage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListIdentityPoolUsage', listIdentityPoolUsage_nextToken - A pagination token for obtaining the next page of results.

$sel:maxResults:ListIdentityPoolUsage', listIdentityPoolUsage_maxResults - The maximum number of results to be returned.

Request Lenses

listIdentityPoolUsage_nextToken :: Lens' ListIdentityPoolUsage (Maybe Text)

A pagination token for obtaining the next page of results.

listIdentityPoolUsage_maxResults :: Lens' ListIdentityPoolUsage (Maybe Int)

The maximum number of results to be returned.

Destructuring the Response

data ListIdentityPoolUsageResponse

Returned for a successful ListIdentityPoolUsage request.

See: newListIdentityPoolUsageResponse smart constructor.

Constructors

ListIdentityPoolUsageResponse'

Fields

identityPoolUsages :: Maybe [IdentityPoolUsage]

Usage information for the identity pools.

nextToken :: Maybe Text

A pagination token for obtaining the next page of results.

maxResults :: Maybe Int

The maximum number of results to be returned.

count :: Maybe Int

Total number of identities for the identity pool.

httpStatus :: Int

The response's http status code.

newListIdentityPoolUsageResponse

Arguments

:: Int
$sel:httpStatus:ListIdentityPoolUsageResponse'

-> ListIdentityPoolUsageResponse

Create a value of ListIdentityPoolUsageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identityPoolUsages:ListIdentityPoolUsageResponse', listIdentityPoolUsageResponse_identityPoolUsages - Usage information for the identity pools.

$sel:nextToken:ListIdentityPoolUsage', listIdentityPoolUsageResponse_nextToken - A pagination token for obtaining the next page of results.

$sel:maxResults:ListIdentityPoolUsage', listIdentityPoolUsageResponse_maxResults - The maximum number of results to be returned.

$sel:count:ListIdentityPoolUsageResponse', listIdentityPoolUsageResponse_count - Total number of identities for the identity pool.

$sel:httpStatus:ListIdentityPoolUsageResponse', listIdentityPoolUsageResponse_httpStatus - The response's http status code.

Response Lenses

listIdentityPoolUsageResponse_identityPoolUsages :: Lens' ListIdentityPoolUsageResponse (Maybe [IdentityPoolUsage])

Usage information for the identity pools.

listIdentityPoolUsageResponse_nextToken :: Lens' ListIdentityPoolUsageResponse (Maybe Text)

A pagination token for obtaining the next page of results.

listIdentityPoolUsageResponse_maxResults :: Lens' ListIdentityPoolUsageResponse (Maybe Int)

The maximum number of results to be returned.

listIdentityPoolUsageResponse_count :: Lens' ListIdentityPoolUsageResponse (Maybe Int)

Total number of identities for the identity pool.

listIdentityPoolUsageResponse_httpStatus :: Lens' ListIdentityPoolUsageResponse Int

The response's http status code.