Network.AWS.CognitoSync.ListRecords

Description

Gets paginated records, optionally changed after a particular sync count for a dataset and identity. With Amazon Cognito Sync, each identity has access only to its own data. Thus, the credentials used to make this API call need to have access to the identity data.

ListRecords can be called with temporary user credentials provided by Cognito Identity or with developer credentials. You should use Cognito Identity credentials to make this API call.

Creating a Request

data ListRecords

A request for a list of records.

See: newListRecords smart constructor.

Constructors

ListRecords'

Fields

nextToken :: Maybe Text

A pagination token for obtaining the next page of results.

maxResults :: Maybe Int

The maximum number of results to be returned.

lastSyncCount :: Maybe Integer

The last server sync count for this record.

syncSessionToken :: Maybe Text

A token containing a session ID, identity ID, and expiration.

identityPoolId :: Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

identityId :: Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

datasetName :: Text

A string of up to 128 characters. Allowed characters are a-z, A-Z, 0-9, '_' (underscore), '-' (dash), and '.' (dot).

newListRecords

Arguments

:: Text
$sel:identityPoolId:ListRecords'

-> Text
$sel:identityId:ListRecords'

-> Text
$sel:datasetName:ListRecords'

-> ListRecords

Create a value of ListRecords with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListRecords', listRecords_nextToken - A pagination token for obtaining the next page of results.

$sel:maxResults:ListRecords', listRecords_maxResults - The maximum number of results to be returned.

$sel:lastSyncCount:ListRecords', listRecords_lastSyncCount - The last server sync count for this record.

$sel:syncSessionToken:ListRecords', listRecords_syncSessionToken - A token containing a session ID, identity ID, and expiration.

$sel:identityPoolId:ListRecords', listRecords_identityPoolId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

$sel:identityId:ListRecords', listRecords_identityId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

$sel:datasetName:ListRecords', listRecords_datasetName - A string of up to 128 characters. Allowed characters are a-z, A-Z, 0-9, '_' (underscore), '-' (dash), and '.' (dot).

Request Lenses

listRecords_nextToken :: Lens' ListRecords (Maybe Text)

A pagination token for obtaining the next page of results.

listRecords_maxResults :: Lens' ListRecords (Maybe Int)

The maximum number of results to be returned.

listRecords_lastSyncCount :: Lens' ListRecords (Maybe Integer)

The last server sync count for this record.

listRecords_syncSessionToken :: Lens' ListRecords (Maybe Text)

A token containing a session ID, identity ID, and expiration.

listRecords_identityPoolId :: Lens' ListRecords Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

listRecords_identityId :: Lens' ListRecords Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

listRecords_datasetName :: Lens' ListRecords Text

A string of up to 128 characters. Allowed characters are a-z, A-Z, 0-9, '_' (underscore), '-' (dash), and '.' (dot).

Destructuring the Response

data ListRecordsResponse

Returned for a successful ListRecordsRequest.

See: newListRecordsResponse smart constructor.

Constructors

ListRecordsResponse'

Fields

nextToken :: Maybe Text

A pagination token for obtaining the next page of results.

datasetSyncCount :: Maybe Integer

Server sync count for this dataset.

records :: Maybe [Record]

A list of all records.

datasetDeletedAfterRequestedSyncCount :: Maybe Bool

A boolean value specifying whether to delete the dataset locally.

count :: Maybe Int

Total number of records.

lastModifiedBy :: Maybe Text

The user/device that made the last change to this record.

datasetExists :: Maybe Bool

Indicates whether the dataset exists.

syncSessionToken :: Maybe Text

A token containing a session ID, identity ID, and expiration.

mergedDatasetNames :: Maybe [Text]

Names of merged datasets.

httpStatus :: Int

The response's http status code.

newListRecordsResponse

Arguments

:: Int
$sel:httpStatus:ListRecordsResponse'

-> ListRecordsResponse

Create a value of ListRecordsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListRecords', listRecordsResponse_nextToken - A pagination token for obtaining the next page of results.

$sel:datasetSyncCount:ListRecordsResponse', listRecordsResponse_datasetSyncCount - Server sync count for this dataset.

$sel:records:ListRecordsResponse', listRecordsResponse_records - A list of all records.

$sel:datasetDeletedAfterRequestedSyncCount:ListRecordsResponse', listRecordsResponse_datasetDeletedAfterRequestedSyncCount - A boolean value specifying whether to delete the dataset locally.

$sel:count:ListRecordsResponse', listRecordsResponse_count - Total number of records.

$sel:lastModifiedBy:ListRecordsResponse', listRecordsResponse_lastModifiedBy - The user/device that made the last change to this record.

$sel:datasetExists:ListRecordsResponse', listRecordsResponse_datasetExists - Indicates whether the dataset exists.

$sel:syncSessionToken:ListRecords', listRecordsResponse_syncSessionToken - A token containing a session ID, identity ID, and expiration.

$sel:mergedDatasetNames:ListRecordsResponse', listRecordsResponse_mergedDatasetNames - Names of merged datasets.

$sel:httpStatus:ListRecordsResponse', listRecordsResponse_httpStatus - The response's http status code.

Response Lenses

listRecordsResponse_nextToken :: Lens' ListRecordsResponse (Maybe Text)

A pagination token for obtaining the next page of results.

listRecordsResponse_datasetSyncCount :: Lens' ListRecordsResponse (Maybe Integer)

Server sync count for this dataset.

listRecordsResponse_records :: Lens' ListRecordsResponse (Maybe [Record])

A list of all records.

listRecordsResponse_datasetDeletedAfterRequestedSyncCount :: Lens' ListRecordsResponse (Maybe Bool)

A boolean value specifying whether to delete the dataset locally.

listRecordsResponse_count :: Lens' ListRecordsResponse (Maybe Int)

Total number of records.

listRecordsResponse_lastModifiedBy :: Lens' ListRecordsResponse (Maybe Text)

The user/device that made the last change to this record.

listRecordsResponse_datasetExists :: Lens' ListRecordsResponse (Maybe Bool)

Indicates whether the dataset exists.

listRecordsResponse_syncSessionToken :: Lens' ListRecordsResponse (Maybe Text)

A token containing a session ID, identity ID, and expiration.

listRecordsResponse_mergedDatasetNames :: Lens' ListRecordsResponse (Maybe [Text])

Names of merged datasets.

listRecordsResponse_httpStatus :: Lens' ListRecordsResponse Int

The response's http status code.