Network.AWS.CognitoSync.RegisterDevice

Description

Registers a device to receive push sync notifications.

This API can only be called with temporary credentials provided by Cognito Identity. You cannot call this API with developer credentials.

Creating a Request

data RegisterDevice

A request to RegisterDevice.

See: newRegisterDevice smart constructor.

Constructors

RegisterDevice'

Fields

identityPoolId :: Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. Here, the ID of the pool that the identity belongs to.

identityId :: Text

The unique ID for this identity.

platform :: Platform

The SNS platform type (e.g. GCM, SDM, APNS, APNS_SANDBOX).

token :: Text

The push token.

newRegisterDevice

Arguments

:: Text
$sel:identityPoolId:RegisterDevice'

-> Text
$sel:identityId:RegisterDevice'

-> Platform
$sel:platform:RegisterDevice'

-> Text
$sel:token:RegisterDevice'

-> RegisterDevice

Create a value of RegisterDevice with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identityPoolId:RegisterDevice', registerDevice_identityPoolId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. Here, the ID of the pool that the identity belongs to.

$sel:identityId:RegisterDevice', registerDevice_identityId - The unique ID for this identity.

$sel:platform:RegisterDevice', registerDevice_platform - The SNS platform type (e.g. GCM, SDM, APNS, APNS_SANDBOX).

$sel:token:RegisterDevice', registerDevice_token - The push token.

Request Lenses

registerDevice_identityPoolId :: Lens' RegisterDevice Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. Here, the ID of the pool that the identity belongs to.

registerDevice_identityId :: Lens' RegisterDevice Text

The unique ID for this identity.

registerDevice_platform :: Lens' RegisterDevice Platform

The SNS platform type (e.g. GCM, SDM, APNS, APNS_SANDBOX).

registerDevice_token :: Lens' RegisterDevice Text

The push token.

Destructuring the Response

data RegisterDeviceResponse

Response to a RegisterDevice request.

See: newRegisterDeviceResponse smart constructor.

Constructors

RegisterDeviceResponse'

Fields

deviceId :: Maybe Text

The unique ID generated for this device by Cognito.

httpStatus :: Int

The response's http status code.

newRegisterDeviceResponse

Arguments

:: Int
$sel:httpStatus:RegisterDeviceResponse'

-> RegisterDeviceResponse

Create a value of RegisterDeviceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceId:RegisterDeviceResponse', registerDeviceResponse_deviceId - The unique ID generated for this device by Cognito.

$sel:httpStatus:RegisterDeviceResponse', registerDeviceResponse_httpStatus - The response's http status code.

Response Lenses

registerDeviceResponse_deviceId :: Lens' RegisterDeviceResponse (Maybe Text)

The unique ID generated for this device by Cognito.

registerDeviceResponse_httpStatus :: Lens' RegisterDeviceResponse Int

The response's http status code.