Network.AWS.CognitoSync.Types

Description

Service Configuration

defaultService :: Service

API version 2014-06-30 of the Amazon Cognito Sync SDK configuration.

Errors

_InvalidConfigurationException :: AsError a => Getting (First ServiceError) a ServiceError

Prism for InvalidConfigurationException' errors.

_ResourceConflictException :: AsError a => Getting (First ServiceError) a ServiceError

Thrown if an update can't be applied because the resource was changed by another call and this would result in a conflict.

_InternalErrorException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates an internal service error.

_ConcurrentModificationException :: AsError a => Getting (First ServiceError) a ServiceError

Thrown if there are parallel requests to modify a resource.

_InvalidParameterException :: AsError a => Getting (First ServiceError) a ServiceError

Thrown when a request parameter does not comply with the associated constraints.

_AlreadyStreamedException :: AsError a => Getting (First ServiceError) a ServiceError

An exception thrown when a bulk publish operation is requested less than 24 hours after a previous bulk publish operation completed successfully.

_InvalidLambdaFunctionOutputException :: AsError a => Getting (First ServiceError) a ServiceError

The AWS Lambda function returned invalid output or an exception.

_LambdaThrottledException :: AsError a => Getting (First ServiceError) a ServiceError

AWS Lambda throttled your account, please contact AWS Support

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Thrown when the limit on the number of objects or operations has been exceeded.

_DuplicateRequestException :: AsError a => Getting (First ServiceError) a ServiceError

An exception thrown when there is an IN_PROGRESS bulk publish operation for the given identity pool.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

Thrown if the resource doesn't exist.

_NotAuthorizedException :: AsError a => Getting (First ServiceError) a ServiceError

Thrown when a user is not authorized to access the requested resource.

_TooManyRequestsException :: AsError a => Getting (First ServiceError) a ServiceError

Thrown if the request is throttled.

BulkPublishStatus

newtype BulkPublishStatus

Constructors

BulkPublishStatus'

Fields

fromBulkPublishStatus :: Text

Bundled Patterns

pattern BulkPublishStatus_FAILED :: BulkPublishStatus

pattern BulkPublishStatus_IN_PROGRESS :: BulkPublishStatus

pattern BulkPublishStatus_NOT_STARTED :: BulkPublishStatus

pattern BulkPublishStatus_SUCCEEDED :: BulkPublishStatus

Operation

newtype Operation

Constructors

Operation'

Fields

fromOperation :: Text

Bundled Patterns

pattern Operation_Remove :: Operation

pattern Operation_Replace :: Operation

Platform

newtype Platform

Constructors

Platform'

Fields

fromPlatform :: Text

Bundled Patterns

pattern Platform_ADM :: Platform

pattern Platform_APNS :: Platform

pattern Platform_APNS_SANDBOX :: Platform

pattern Platform_GCM :: Platform

StreamingStatus

newtype StreamingStatus

Constructors

StreamingStatus'

Fields

fromStreamingStatus :: Text

Bundled Patterns

pattern StreamingStatus_DISABLED :: StreamingStatus

pattern StreamingStatus_ENABLED :: StreamingStatus

CognitoStreams

data CognitoStreams

Configuration options for configure Cognito streams.

See: newCognitoStreams smart constructor.

Constructors

CognitoStreams'

Fields

roleArn :: Maybe Text

The ARN of the role Amazon Cognito can assume in order to publish to the stream. This role must grant access to Amazon Cognito (cognito-sync) to invoke PutRecord on your Cognito stream.

streamName :: Maybe Text

The name of the Cognito stream to receive updates. This stream must be in the developers account and in the same region as the identity pool.

streamingStatus :: Maybe StreamingStatus

Status of the Cognito streams. Valid values are:

ENABLED - Streaming of updates to identity pool is enabled.

DISABLED - Streaming of updates to identity pool is disabled. Bulk publish will also fail if StreamingStatus is DISABLED.

newCognitoStreams :: CognitoStreams

Create a value of CognitoStreams with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:roleArn:CognitoStreams', cognitoStreams_roleArn - The ARN of the role Amazon Cognito can assume in order to publish to the stream. This role must grant access to Amazon Cognito (cognito-sync) to invoke PutRecord on your Cognito stream.

$sel:streamName:CognitoStreams', cognitoStreams_streamName - The name of the Cognito stream to receive updates. This stream must be in the developers account and in the same region as the identity pool.

$sel:streamingStatus:CognitoStreams', cognitoStreams_streamingStatus - Status of the Cognito streams. Valid values are:

ENABLED - Streaming of updates to identity pool is enabled.

DISABLED - Streaming of updates to identity pool is disabled. Bulk publish will also fail if StreamingStatus is DISABLED.

cognitoStreams_roleArn :: Lens' CognitoStreams (Maybe Text)

The ARN of the role Amazon Cognito can assume in order to publish to the stream. This role must grant access to Amazon Cognito (cognito-sync) to invoke PutRecord on your Cognito stream.

cognitoStreams_streamName :: Lens' CognitoStreams (Maybe Text)

The name of the Cognito stream to receive updates. This stream must be in the developers account and in the same region as the identity pool.

cognitoStreams_streamingStatus :: Lens' CognitoStreams (Maybe StreamingStatus)

Status of the Cognito streams. Valid values are:

ENABLED - Streaming of updates to identity pool is enabled.

DISABLED - Streaming of updates to identity pool is disabled. Bulk publish will also fail if StreamingStatus is DISABLED.

Dataset

data Dataset

A collection of data for an identity pool. An identity pool can have multiple datasets. A dataset is per identity and can be general or associated with a particular entity in an application (like a saved game). Datasets are automatically created if they don't exist. Data is synced by dataset, and a dataset can hold up to 1MB of key-value pairs.

See: newDataset smart constructor.

Constructors

Dataset'

Fields

lastModifiedDate :: Maybe POSIX

Date when the dataset was last modified.

numRecords :: Maybe Integer

Number of records in this dataset.

creationDate :: Maybe POSIX

Date on which the dataset was created.

dataStorage :: Maybe Integer

Total size in bytes of the records in this dataset.

identityId :: Maybe Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

lastModifiedBy :: Maybe Text

The device that made the last change to this dataset.

datasetName :: Maybe Text

A string of up to 128 characters. Allowed characters are a-z, A-Z, 0-9, '_' (underscore), '-' (dash), and '.' (dot).

newDataset :: Dataset

Create a value of Dataset with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedDate:Dataset', dataset_lastModifiedDate - Date when the dataset was last modified.

$sel:numRecords:Dataset', dataset_numRecords - Number of records in this dataset.

$sel:creationDate:Dataset', dataset_creationDate - Date on which the dataset was created.

$sel:dataStorage:Dataset', dataset_dataStorage - Total size in bytes of the records in this dataset.

$sel:identityId:Dataset', dataset_identityId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

$sel:lastModifiedBy:Dataset', dataset_lastModifiedBy - The device that made the last change to this dataset.

$sel:datasetName:Dataset', dataset_datasetName - A string of up to 128 characters. Allowed characters are a-z, A-Z, 0-9, '_' (underscore), '-' (dash), and '.' (dot).

dataset_lastModifiedDate :: Lens' Dataset (Maybe UTCTime)

Date when the dataset was last modified.

dataset_numRecords :: Lens' Dataset (Maybe Integer)

Number of records in this dataset.

dataset_creationDate :: Lens' Dataset (Maybe UTCTime)

Date on which the dataset was created.

dataset_dataStorage :: Lens' Dataset (Maybe Integer)

Total size in bytes of the records in this dataset.

dataset_identityId :: Lens' Dataset (Maybe Text)

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

dataset_lastModifiedBy :: Lens' Dataset (Maybe Text)

The device that made the last change to this dataset.

dataset_datasetName :: Lens' Dataset (Maybe Text)

A string of up to 128 characters. Allowed characters are a-z, A-Z, 0-9, '_' (underscore), '-' (dash), and '.' (dot).

IdentityPoolUsage

data IdentityPoolUsage

Usage information for the identity pool.

See: newIdentityPoolUsage smart constructor.

Constructors

IdentityPoolUsage'

Fields

lastModifiedDate :: Maybe POSIX

Date on which the identity pool was last modified.

identityPoolId :: Maybe Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

syncSessionsCount :: Maybe Integer

Number of sync sessions for the identity pool.

dataStorage :: Maybe Integer

Data storage information for the identity pool.

newIdentityPoolUsage :: IdentityPoolUsage

Create a value of IdentityPoolUsage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedDate:IdentityPoolUsage', identityPoolUsage_lastModifiedDate - Date on which the identity pool was last modified.

$sel:identityPoolId:IdentityPoolUsage', identityPoolUsage_identityPoolId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

$sel:syncSessionsCount:IdentityPoolUsage', identityPoolUsage_syncSessionsCount - Number of sync sessions for the identity pool.

$sel:dataStorage:IdentityPoolUsage', identityPoolUsage_dataStorage - Data storage information for the identity pool.

identityPoolUsage_lastModifiedDate :: Lens' IdentityPoolUsage (Maybe UTCTime)

Date on which the identity pool was last modified.

identityPoolUsage_identityPoolId :: Lens' IdentityPoolUsage (Maybe Text)

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

identityPoolUsage_syncSessionsCount :: Lens' IdentityPoolUsage (Maybe Integer)

Number of sync sessions for the identity pool.

identityPoolUsage_dataStorage :: Lens' IdentityPoolUsage (Maybe Integer)

Data storage information for the identity pool.

IdentityUsage

data IdentityUsage

Usage information for the identity.

See: newIdentityUsage smart constructor.

Constructors

IdentityUsage'

Fields

lastModifiedDate :: Maybe POSIX

Date on which the identity was last modified.

identityPoolId :: Maybe Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

datasetCount :: Maybe Int

Number of datasets for the identity.

dataStorage :: Maybe Integer

Total data storage for this identity.

identityId :: Maybe Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

newIdentityUsage :: IdentityUsage

Create a value of IdentityUsage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedDate:IdentityUsage', identityUsage_lastModifiedDate - Date on which the identity was last modified.

$sel:identityPoolId:IdentityUsage', identityUsage_identityPoolId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

$sel:datasetCount:IdentityUsage', identityUsage_datasetCount - Number of datasets for the identity.

$sel:dataStorage:IdentityUsage', identityUsage_dataStorage - Total data storage for this identity.

$sel:identityId:IdentityUsage', identityUsage_identityId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

identityUsage_lastModifiedDate :: Lens' IdentityUsage (Maybe UTCTime)

Date on which the identity was last modified.

identityUsage_identityPoolId :: Lens' IdentityUsage (Maybe Text)

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

identityUsage_datasetCount :: Lens' IdentityUsage (Maybe Int)

Number of datasets for the identity.

identityUsage_dataStorage :: Lens' IdentityUsage (Maybe Integer)

Total data storage for this identity.

identityUsage_identityId :: Lens' IdentityUsage (Maybe Text)

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

PushSync

data PushSync

Configuration options to be applied to the identity pool.

See: newPushSync smart constructor.

Constructors

PushSync'

Fields

roleArn :: Maybe Text

A role configured to allow Cognito to call SNS on behalf of the developer.

applicationArns :: Maybe [Text]

List of SNS platform application ARNs that could be used by clients.

newPushSync :: PushSync

Create a value of PushSync with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:roleArn:PushSync', pushSync_roleArn - A role configured to allow Cognito to call SNS on behalf of the developer.

$sel:applicationArns:PushSync', pushSync_applicationArns - List of SNS platform application ARNs that could be used by clients.

pushSync_roleArn :: Lens' PushSync (Maybe Text)

A role configured to allow Cognito to call SNS on behalf of the developer.

pushSync_applicationArns :: Lens' PushSync (Maybe [Text])

List of SNS platform application ARNs that could be used by clients.

Record

data Record

The basic data structure of a dataset.

See: newRecord smart constructor.

Constructors

Record'

Fields

key :: Maybe Text

The key for the record.

deviceLastModifiedDate :: Maybe POSIX

The last modified date of the client device.

lastModifiedDate :: Maybe POSIX

The date on which the record was last modified.

syncCount :: Maybe Integer

The server sync count for this record.

value :: Maybe Text

The value for the record.

lastModifiedBy :: Maybe Text

The user/device that made the last change to this record.

newRecord :: Record

Create a value of Record with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:Record', record_key - The key for the record.

$sel:deviceLastModifiedDate:Record', record_deviceLastModifiedDate - The last modified date of the client device.

$sel:lastModifiedDate:Record', record_lastModifiedDate - The date on which the record was last modified.

$sel:syncCount:Record', record_syncCount - The server sync count for this record.

$sel:value:Record', record_value - The value for the record.

$sel:lastModifiedBy:Record', record_lastModifiedBy - The user/device that made the last change to this record.

record_key :: Lens' Record (Maybe Text)

The key for the record.

record_deviceLastModifiedDate :: Lens' Record (Maybe UTCTime)

The last modified date of the client device.

record_lastModifiedDate :: Lens' Record (Maybe UTCTime)

The date on which the record was last modified.

record_syncCount :: Lens' Record (Maybe Integer)

The server sync count for this record.

record_value :: Lens' Record (Maybe Text)

The value for the record.

record_lastModifiedBy :: Lens' Record (Maybe Text)

The user/device that made the last change to this record.

RecordPatch

data RecordPatch

An update operation for a record.

See: newRecordPatch smart constructor.

Constructors

RecordPatch'

Fields

deviceLastModifiedDate :: Maybe POSIX

The last modified date of the client device.

value :: Maybe Text

The value associated with the record patch.

op :: Operation

An operation, either replace or remove.

key :: Text

The key associated with the record patch.

syncCount :: Integer

Last known server sync count for this record. Set to 0 if unknown.

newRecordPatch

Arguments

:: Operation
$sel:op:RecordPatch'

-> Text
$sel:key:RecordPatch'

-> Integer
$sel:syncCount:RecordPatch'

-> RecordPatch

Create a value of RecordPatch with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceLastModifiedDate:RecordPatch', recordPatch_deviceLastModifiedDate - The last modified date of the client device.

$sel:value:RecordPatch', recordPatch_value - The value associated with the record patch.

$sel:op:RecordPatch', recordPatch_op - An operation, either replace or remove.

$sel:key:RecordPatch', recordPatch_key - The key associated with the record patch.

$sel:syncCount:RecordPatch', recordPatch_syncCount - Last known server sync count for this record. Set to 0 if unknown.

recordPatch_deviceLastModifiedDate :: Lens' RecordPatch (Maybe UTCTime)

The last modified date of the client device.

recordPatch_value :: Lens' RecordPatch (Maybe Text)

The value associated with the record patch.

recordPatch_op :: Lens' RecordPatch Operation

An operation, either replace or remove.

recordPatch_key :: Lens' RecordPatch Text

The key associated with the record patch.

recordPatch_syncCount :: Lens' RecordPatch Integer

Last known server sync count for this record. Set to 0 if unknown.