Network.AWS.CognitoSync.Types.Record

Description

Documentation

data Record

The basic data structure of a dataset.

See: newRecord smart constructor.

Constructors

Record'

Fields

key :: Maybe Text

The key for the record.

deviceLastModifiedDate :: Maybe POSIX

The last modified date of the client device.

lastModifiedDate :: Maybe POSIX

The date on which the record was last modified.

syncCount :: Maybe Integer

The server sync count for this record.

value :: Maybe Text

The value for the record.

lastModifiedBy :: Maybe Text

The user/device that made the last change to this record.

newRecord :: Record

Create a value of Record with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:Record', record_key - The key for the record.

$sel:deviceLastModifiedDate:Record', record_deviceLastModifiedDate - The last modified date of the client device.

$sel:lastModifiedDate:Record', record_lastModifiedDate - The date on which the record was last modified.

$sel:syncCount:Record', record_syncCount - The server sync count for this record.

$sel:value:Record', record_value - The value for the record.

$sel:lastModifiedBy:Record', record_lastModifiedBy - The user/device that made the last change to this record.

record_key :: Lens' Record (Maybe Text)

The key for the record.

record_deviceLastModifiedDate :: Lens' Record (Maybe UTCTime)

The last modified date of the client device.

record_lastModifiedDate :: Lens' Record (Maybe UTCTime)

The date on which the record was last modified.

record_syncCount :: Lens' Record (Maybe Integer)

The server sync count for this record.

record_value :: Lens' Record (Maybe Text)

The value for the record.

record_lastModifiedBy :: Lens' Record (Maybe Text)

The user/device that made the last change to this record.