Network.AWS.CognitoSync.UnsubscribeFromDataset

Description

Unsubscribes from receiving notifications when a dataset is modified by another device.

This API can only be called with temporary credentials provided by Cognito Identity. You cannot call this API with developer credentials.

Creating a Request

data UnsubscribeFromDataset

A request to UnsubscribeFromDataset.

See: newUnsubscribeFromDataset smart constructor.

Constructors

UnsubscribeFromDataset'

Fields

identityPoolId :: Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. The ID of the pool to which this identity belongs.

identityId :: Text

Unique ID for this identity.

datasetName :: Text

The name of the dataset from which to unsubcribe.

deviceId :: Text

The unique ID generated for this device by Cognito.

newUnsubscribeFromDataset

Arguments

:: Text
$sel:identityPoolId:UnsubscribeFromDataset'

-> Text
$sel:identityId:UnsubscribeFromDataset'

-> Text
$sel:datasetName:UnsubscribeFromDataset'

-> Text
$sel:deviceId:UnsubscribeFromDataset'

-> UnsubscribeFromDataset

Create a value of UnsubscribeFromDataset with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identityPoolId:UnsubscribeFromDataset', unsubscribeFromDataset_identityPoolId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. The ID of the pool to which this identity belongs.

$sel:identityId:UnsubscribeFromDataset', unsubscribeFromDataset_identityId - Unique ID for this identity.

$sel:datasetName:UnsubscribeFromDataset', unsubscribeFromDataset_datasetName - The name of the dataset from which to unsubcribe.

$sel:deviceId:UnsubscribeFromDataset', unsubscribeFromDataset_deviceId - The unique ID generated for this device by Cognito.

Request Lenses

unsubscribeFromDataset_identityPoolId :: Lens' UnsubscribeFromDataset Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. The ID of the pool to which this identity belongs.

unsubscribeFromDataset_identityId :: Lens' UnsubscribeFromDataset Text

Unique ID for this identity.

unsubscribeFromDataset_datasetName :: Lens' UnsubscribeFromDataset Text

The name of the dataset from which to unsubcribe.

unsubscribeFromDataset_deviceId :: Lens' UnsubscribeFromDataset Text

The unique ID generated for this device by Cognito.

Destructuring the Response

data UnsubscribeFromDatasetResponse

Response to an UnsubscribeFromDataset request.

See: newUnsubscribeFromDatasetResponse smart constructor.

Constructors

UnsubscribeFromDatasetResponse'

Fields

httpStatus :: Int

The response's http status code.

newUnsubscribeFromDatasetResponse

Arguments

:: Int
$sel:httpStatus:UnsubscribeFromDatasetResponse'

-> UnsubscribeFromDatasetResponse

Create a value of UnsubscribeFromDatasetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UnsubscribeFromDatasetResponse', unsubscribeFromDatasetResponse_httpStatus - The response's http status code.

Response Lenses

unsubscribeFromDatasetResponse_httpStatus :: Lens' UnsubscribeFromDatasetResponse Int

The response's http status code.