Network.AWS.CognitoSync.UpdateRecords

Description

Posts updates to records and adds and deletes records for a dataset and user.

The sync count in the record patch is your last known sync count for that record. The server will reject an UpdateRecords request with a ResourceConflictException if you try to patch a record with a new value but a stale sync count.

For example, if the sync count on the server is 5 for a key called highScore and you try and submit a new highScore with sync count of 4, the request will be rejected. To obtain the current sync count for a record, call ListRecords. On a successful update of the record, the response returns the new sync count for that record. You should present that sync count the next time you try to update that same record. When the record does not exist, specify the sync count as 0.

This API can be called with temporary user credentials provided by Cognito Identity or with developer credentials.

Creating a Request

data UpdateRecords

A request to post updates to records or add and delete records for a dataset and user.

See: newUpdateRecords smart constructor.

Constructors

UpdateRecords'

Fields

recordPatches :: Maybe [RecordPatch]

A list of patch operations.

deviceId :: Maybe Text

The unique ID generated for this device by Cognito.

clientContext :: Maybe Text

Intended to supply a device ID that will populate the lastModifiedBy field referenced in other methods. The ClientContext field is not yet implemented.

identityPoolId :: Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

identityId :: Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

datasetName :: Text

A string of up to 128 characters. Allowed characters are a-z, A-Z, 0-9, '_' (underscore), '-' (dash), and '.' (dot).

syncSessionToken :: Text

The SyncSessionToken returned by a previous call to ListRecords for this dataset and identity.

newUpdateRecords

Arguments

:: Text
$sel:identityPoolId:UpdateRecords'

-> Text
$sel:identityId:UpdateRecords'

-> Text
$sel:datasetName:UpdateRecords'

-> Text
$sel:syncSessionToken:UpdateRecords'

-> UpdateRecords

Create a value of UpdateRecords with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:recordPatches:UpdateRecords', updateRecords_recordPatches - A list of patch operations.

$sel:deviceId:UpdateRecords', updateRecords_deviceId - The unique ID generated for this device by Cognito.

$sel:clientContext:UpdateRecords', updateRecords_clientContext - Intended to supply a device ID that will populate the lastModifiedBy field referenced in other methods. The ClientContext field is not yet implemented.

$sel:identityPoolId:UpdateRecords', updateRecords_identityPoolId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

$sel:identityId:UpdateRecords', updateRecords_identityId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

$sel:datasetName:UpdateRecords', updateRecords_datasetName - A string of up to 128 characters. Allowed characters are a-z, A-Z, 0-9, '_' (underscore), '-' (dash), and '.' (dot).

$sel:syncSessionToken:UpdateRecords', updateRecords_syncSessionToken - The SyncSessionToken returned by a previous call to ListRecords for this dataset and identity.

Request Lenses

updateRecords_recordPatches :: Lens' UpdateRecords (Maybe [RecordPatch])

A list of patch operations.

updateRecords_deviceId :: Lens' UpdateRecords (Maybe Text)

The unique ID generated for this device by Cognito.

updateRecords_clientContext :: Lens' UpdateRecords (Maybe Text)

Intended to supply a device ID that will populate the lastModifiedBy field referenced in other methods. The ClientContext field is not yet implemented.

updateRecords_identityPoolId :: Lens' UpdateRecords Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

updateRecords_identityId :: Lens' UpdateRecords Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

updateRecords_datasetName :: Lens' UpdateRecords Text

A string of up to 128 characters. Allowed characters are a-z, A-Z, 0-9, '_' (underscore), '-' (dash), and '.' (dot).

updateRecords_syncSessionToken :: Lens' UpdateRecords Text

The SyncSessionToken returned by a previous call to ListRecords for this dataset and identity.

Destructuring the Response

data UpdateRecordsResponse

Returned for a successful UpdateRecordsRequest.

See: newUpdateRecordsResponse smart constructor.

Constructors

UpdateRecordsResponse'

Fields

records :: Maybe [Record]

A list of records that have been updated.

httpStatus :: Int

The response's http status code.

newUpdateRecordsResponse

Arguments

:: Int
$sel:httpStatus:UpdateRecordsResponse'

-> UpdateRecordsResponse

Create a value of UpdateRecordsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:records:UpdateRecordsResponse', updateRecordsResponse_records - A list of records that have been updated.

$sel:httpStatus:UpdateRecordsResponse', updateRecordsResponse_httpStatus - The response's http status code.

Response Lenses

updateRecordsResponse_records :: Lens' UpdateRecordsResponse (Maybe [Record])

A list of records that have been updated.

updateRecordsResponse_httpStatus :: Lens' UpdateRecordsResponse Int

The response's http status code.