Network.AWS.Comprehend.BatchDetectDominantLanguage

Description

Determines the dominant language of the input text for a batch of documents. For a list of languages that Amazon Comprehend can detect, see Amazon Comprehend Supported Languages.

Creating a Request

data BatchDetectDominantLanguage

See: newBatchDetectDominantLanguage smart constructor.

Constructors

BatchDetectDominantLanguage'

Fields

textList :: Sensitive [Sensitive Text]

A list containing the text of the input documents. The list can contain a maximum of 25 documents. Each document should contain at least 20 characters and must contain fewer than 5,000 bytes of UTF-8 encoded characters.

newBatchDetectDominantLanguage :: BatchDetectDominantLanguage

Create a value of BatchDetectDominantLanguage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:textList:BatchDetectDominantLanguage', batchDetectDominantLanguage_textList - A list containing the text of the input documents. The list can contain a maximum of 25 documents. Each document should contain at least 20 characters and must contain fewer than 5,000 bytes of UTF-8 encoded characters.

Request Lenses

batchDetectDominantLanguage_textList :: Lens' BatchDetectDominantLanguage [Text]

A list containing the text of the input documents. The list can contain a maximum of 25 documents. Each document should contain at least 20 characters and must contain fewer than 5,000 bytes of UTF-8 encoded characters.

Destructuring the Response

data BatchDetectDominantLanguageResponse

See: newBatchDetectDominantLanguageResponse smart constructor.

Constructors

BatchDetectDominantLanguageResponse'

Fields

httpStatus :: Int

The response's http status code.

resultList :: [BatchDetectDominantLanguageItemResult]

A list of objects containing the results of the operation. The results are sorted in ascending order by the Index field and match the order of the documents in the input list. If all of the documents contain an error, the ResultList is empty.

errorList :: [BatchItemError]

A list containing one object for each document that contained an error. The results are sorted in ascending order by the Index field and match the order of the documents in the input list. If there are no errors in the batch, the ErrorList is empty.

newBatchDetectDominantLanguageResponse

Arguments

:: Int
$sel:httpStatus:BatchDetectDominantLanguageResponse'

-> BatchDetectDominantLanguageResponse

Create a value of BatchDetectDominantLanguageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:BatchDetectDominantLanguageResponse', batchDetectDominantLanguageResponse_httpStatus - The response's http status code.

$sel:resultList:BatchDetectDominantLanguageResponse', batchDetectDominantLanguageResponse_resultList - A list of objects containing the results of the operation. The results are sorted in ascending order by the Index field and match the order of the documents in the input list. If all of the documents contain an error, the ResultList is empty.

$sel:errorList:BatchDetectDominantLanguageResponse', batchDetectDominantLanguageResponse_errorList - A list containing one object for each document that contained an error. The results are sorted in ascending order by the Index field and match the order of the documents in the input list. If there are no errors in the batch, the ErrorList is empty.

Response Lenses

batchDetectDominantLanguageResponse_httpStatus :: Lens' BatchDetectDominantLanguageResponse Int

The response's http status code.

batchDetectDominantLanguageResponse_resultList :: Lens' BatchDetectDominantLanguageResponse [BatchDetectDominantLanguageItemResult]

A list of objects containing the results of the operation. The results are sorted in ascending order by the Index field and match the order of the documents in the input list. If all of the documents contain an error, the ResultList is empty.

batchDetectDominantLanguageResponse_errorList :: Lens' BatchDetectDominantLanguageResponse [BatchItemError]

A list containing one object for each document that contained an error. The results are sorted in ascending order by the Index field and match the order of the documents in the input list. If there are no errors in the batch, the ErrorList is empty.