Network.AWS.Comprehend.BatchDetectSentiment

Description

Inspects a batch of documents and returns an inference of the prevailing sentiment, POSITIVE, NEUTRAL, MIXED, or NEGATIVE, in each one.

Creating a Request

data BatchDetectSentiment

See: newBatchDetectSentiment smart constructor.

Constructors

BatchDetectSentiment'

Fields

textList :: Sensitive [Sensitive Text]

A list containing the text of the input documents. The list can contain a maximum of 25 documents. Each document must contain fewer that 5,000 bytes of UTF-8 encoded characters.

languageCode :: LanguageCode

The language of the input documents. You can specify any of the primary languages supported by Amazon Comprehend. All documents must be in the same language.

newBatchDetectSentiment

Arguments

:: LanguageCode
$sel:languageCode:BatchDetectSentiment'

-> BatchDetectSentiment

Create a value of BatchDetectSentiment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:textList:BatchDetectSentiment', batchDetectSentiment_textList - A list containing the text of the input documents. The list can contain a maximum of 25 documents. Each document must contain fewer that 5,000 bytes of UTF-8 encoded characters.

$sel:languageCode:BatchDetectSentiment', batchDetectSentiment_languageCode - The language of the input documents. You can specify any of the primary languages supported by Amazon Comprehend. All documents must be in the same language.

Request Lenses

batchDetectSentiment_textList :: Lens' BatchDetectSentiment [Text]

A list containing the text of the input documents. The list can contain a maximum of 25 documents. Each document must contain fewer that 5,000 bytes of UTF-8 encoded characters.

batchDetectSentiment_languageCode :: Lens' BatchDetectSentiment LanguageCode

The language of the input documents. You can specify any of the primary languages supported by Amazon Comprehend. All documents must be in the same language.

Destructuring the Response

data BatchDetectSentimentResponse

See: newBatchDetectSentimentResponse smart constructor.

Constructors

BatchDetectSentimentResponse'

Fields

httpStatus :: Int

The response's http status code.

resultList :: [BatchDetectSentimentItemResult]

A list of objects containing the results of the operation. The results are sorted in ascending order by the Index field and match the order of the documents in the input list. If all of the documents contain an error, the ResultList is empty.

errorList :: [BatchItemError]

A list containing one object for each document that contained an error. The results are sorted in ascending order by the Index field and match the order of the documents in the input list. If there are no errors in the batch, the ErrorList is empty.

newBatchDetectSentimentResponse

Arguments

:: Int
$sel:httpStatus:BatchDetectSentimentResponse'

-> BatchDetectSentimentResponse

Create a value of BatchDetectSentimentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:BatchDetectSentimentResponse', batchDetectSentimentResponse_httpStatus - The response's http status code.

$sel:resultList:BatchDetectSentimentResponse', batchDetectSentimentResponse_resultList - A list of objects containing the results of the operation. The results are sorted in ascending order by the Index field and match the order of the documents in the input list. If all of the documents contain an error, the ResultList is empty.

$sel:errorList:BatchDetectSentimentResponse', batchDetectSentimentResponse_errorList - A list containing one object for each document that contained an error. The results are sorted in ascending order by the Index field and match the order of the documents in the input list. If there are no errors in the batch, the ErrorList is empty.

Response Lenses

batchDetectSentimentResponse_httpStatus :: Lens' BatchDetectSentimentResponse Int

The response's http status code.

batchDetectSentimentResponse_resultList :: Lens' BatchDetectSentimentResponse [BatchDetectSentimentItemResult]

A list of objects containing the results of the operation. The results are sorted in ascending order by the Index field and match the order of the documents in the input list. If all of the documents contain an error, the ResultList is empty.

batchDetectSentimentResponse_errorList :: Lens' BatchDetectSentimentResponse [BatchItemError]

A list containing one object for each document that contained an error. The results are sorted in ascending order by the Index field and match the order of the documents in the input list. If there are no errors in the batch, the ErrorList is empty.