Network.AWS.Comprehend.ClassifyDocument

Description

Creates a new document classification request to analyze a single document in real-time, using a previously created and trained custom model and an endpoint.

Creating a Request

data ClassifyDocument

See: newClassifyDocument smart constructor.

Constructors

ClassifyDocument'

Fields

text :: Sensitive Text

The document text to be analyzed.

endpointArn :: Text

The Amazon Resource Number (ARN) of the endpoint.

newClassifyDocument

Arguments

:: Text
$sel:text:ClassifyDocument'

-> Text
$sel:endpointArn:ClassifyDocument'

-> ClassifyDocument

Create a value of ClassifyDocument with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:text:ClassifyDocument', classifyDocument_text - The document text to be analyzed.

$sel:endpointArn:ClassifyDocument', classifyDocument_endpointArn - The Amazon Resource Number (ARN) of the endpoint.

Request Lenses

classifyDocument_text :: Lens' ClassifyDocument Text

The document text to be analyzed.

classifyDocument_endpointArn :: Lens' ClassifyDocument Text

The Amazon Resource Number (ARN) of the endpoint.

Destructuring the Response

data ClassifyDocumentResponse

See: newClassifyDocumentResponse smart constructor.

Constructors

ClassifyDocumentResponse'

Fields

labels :: Maybe [DocumentLabel]

The labels used the document being analyzed. These are used for multi-label trained models. Individual labels represent different categories that are related in some manner and are not mutually exclusive. For example, a movie can be just an action movie, or it can be an action movie, a science fiction movie, and a comedy, all at the same time.

classes :: Maybe [DocumentClass]

The classes used by the document being analyzed. These are used for multi-class trained models. Individual classes are mutually exclusive and each document is expected to have only a single class assigned to it. For example, an animal can be a dog or a cat, but not both at the same time.

httpStatus :: Int

The response's http status code.

newClassifyDocumentResponse

Arguments

:: Int
$sel:httpStatus:ClassifyDocumentResponse'

-> ClassifyDocumentResponse

Create a value of ClassifyDocumentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:labels:ClassifyDocumentResponse', classifyDocumentResponse_labels - The labels used the document being analyzed. These are used for multi-label trained models. Individual labels represent different categories that are related in some manner and are not mutually exclusive. For example, a movie can be just an action movie, or it can be an action movie, a science fiction movie, and a comedy, all at the same time.

$sel:classes:ClassifyDocumentResponse', classifyDocumentResponse_classes - The classes used by the document being analyzed. These are used for multi-class trained models. Individual classes are mutually exclusive and each document is expected to have only a single class assigned to it. For example, an animal can be a dog or a cat, but not both at the same time.

$sel:httpStatus:ClassifyDocumentResponse', classifyDocumentResponse_httpStatus - The response's http status code.

Response Lenses

classifyDocumentResponse_labels :: Lens' ClassifyDocumentResponse (Maybe [DocumentLabel])

The labels used the document being analyzed. These are used for multi-label trained models. Individual labels represent different categories that are related in some manner and are not mutually exclusive. For example, a movie can be just an action movie, or it can be an action movie, a science fiction movie, and a comedy, all at the same time.

classifyDocumentResponse_classes :: Lens' ClassifyDocumentResponse (Maybe [DocumentClass])

The classes used by the document being analyzed. These are used for multi-class trained models. Individual classes are mutually exclusive and each document is expected to have only a single class assigned to it. For example, an animal can be a dog or a cat, but not both at the same time.

classifyDocumentResponse_httpStatus :: Lens' ClassifyDocumentResponse Int

The response's http status code.