Network.AWS.Comprehend.CreateDocumentClassifier

Description

Creates a new document classifier that you can use to categorize documents. To create a classifier, you provide a set of training documents that labeled with the categories that you want to use. After the classifier is trained you can use it to categorize a set of labeled documents into the categories. For more information, see how-document-classification.

Creating a Request

data CreateDocumentClassifier

See: newCreateDocumentClassifier smart constructor.

Constructors

CreateDocumentClassifier'

Fields

versionName :: Maybe Text

The version name given to the newly created classifier. Version names can have a maximum of 256 characters. Alphanumeric characters, hyphens (-) and underscores (_) are allowed. The version name must be unique among all models with the same classifier name in the account/AWS Region.

modelKmsKeyId :: Maybe Text

ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt trained custom models. The ModelKmsKeyId can be either of the following formats:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

mode :: Maybe DocumentClassifierMode

Indicates the mode in which the classifier will be trained. The classifier can be trained in multi-class mode, which identifies one and only one class for each document, or multi-label mode, which identifies one or more labels for each document. In multi-label mode, multiple labels for an individual document are separated by a delimiter. The default delimiter between labels is a pipe (|).

vpcConfig :: Maybe VpcConfig

Configuration parameters for an optional private Virtual Private Cloud (VPC) containing the resources you are using for your custom classifier. For more information, see Amazon VPC.

volumeKmsKeyId :: Maybe Text

ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt data on the storage volume attached to the ML compute instance(s) that process the analysis job. The VolumeKmsKeyId can be either of the following formats:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

outputDataConfig :: Maybe DocumentClassifierOutputDataConfig

Enables the addition of output results configuration parameters for custom classifier jobs.

clientRequestToken :: Maybe Text

A unique identifier for the request. If you don't set the client request token, Amazon Comprehend generates one.

tags :: Maybe [Tag]

Tags to be associated with the document classifier being created. A tag is a key-value pair that adds as a metadata to a resource used by Amazon Comprehend. For example, a tag with "Sales" as the key might be added to a resource to indicate its use by the sales department.

documentClassifierName :: Text

The name of the document classifier.

dataAccessRoleArn :: Text

The Amazon Resource Name (ARN) of the AWS Identity and Management (IAM) role that grants Amazon Comprehend read access to your input data.

inputDataConfig :: DocumentClassifierInputDataConfig

Specifies the format and location of the input data for the job.

languageCode :: LanguageCode

The language of the input documents. You can specify any of the following languages supported by Amazon Comprehend: German ("de"), English ("en"), Spanish ("es"), French ("fr"), Italian ("it"), or Portuguese ("pt"). All documents must be in the same language.

newCreateDocumentClassifier

Arguments

:: Text
$sel:documentClassifierName:CreateDocumentClassifier'

-> Text
$sel:dataAccessRoleArn:CreateDocumentClassifier'

-> DocumentClassifierInputDataConfig
$sel:inputDataConfig:CreateDocumentClassifier'

-> LanguageCode
$sel:languageCode:CreateDocumentClassifier'

-> CreateDocumentClassifier

Create a value of CreateDocumentClassifier with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:versionName:CreateDocumentClassifier', createDocumentClassifier_versionName - The version name given to the newly created classifier. Version names can have a maximum of 256 characters. Alphanumeric characters, hyphens (-) and underscores (_) are allowed. The version name must be unique among all models with the same classifier name in the account/AWS Region.

$sel:modelKmsKeyId:CreateDocumentClassifier', createDocumentClassifier_modelKmsKeyId - ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt trained custom models. The ModelKmsKeyId can be either of the following formats:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

$sel:mode:CreateDocumentClassifier', createDocumentClassifier_mode - Indicates the mode in which the classifier will be trained. The classifier can be trained in multi-class mode, which identifies one and only one class for each document, or multi-label mode, which identifies one or more labels for each document. In multi-label mode, multiple labels for an individual document are separated by a delimiter. The default delimiter between labels is a pipe (|).

$sel:vpcConfig:CreateDocumentClassifier', createDocumentClassifier_vpcConfig - Configuration parameters for an optional private Virtual Private Cloud (VPC) containing the resources you are using for your custom classifier. For more information, see Amazon VPC.

$sel:volumeKmsKeyId:CreateDocumentClassifier', createDocumentClassifier_volumeKmsKeyId - ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt data on the storage volume attached to the ML compute instance(s) that process the analysis job. The VolumeKmsKeyId can be either of the following formats:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

$sel:outputDataConfig:CreateDocumentClassifier', createDocumentClassifier_outputDataConfig - Enables the addition of output results configuration parameters for custom classifier jobs.

$sel:clientRequestToken:CreateDocumentClassifier', createDocumentClassifier_clientRequestToken - A unique identifier for the request. If you don't set the client request token, Amazon Comprehend generates one.

$sel:tags:CreateDocumentClassifier', createDocumentClassifier_tags - Tags to be associated with the document classifier being created. A tag is a key-value pair that adds as a metadata to a resource used by Amazon Comprehend. For example, a tag with "Sales" as the key might be added to a resource to indicate its use by the sales department.

$sel:documentClassifierName:CreateDocumentClassifier', createDocumentClassifier_documentClassifierName - The name of the document classifier.

$sel:dataAccessRoleArn:CreateDocumentClassifier', createDocumentClassifier_dataAccessRoleArn - The Amazon Resource Name (ARN) of the AWS Identity and Management (IAM) role that grants Amazon Comprehend read access to your input data.

$sel:inputDataConfig:CreateDocumentClassifier', createDocumentClassifier_inputDataConfig - Specifies the format and location of the input data for the job.

$sel:languageCode:CreateDocumentClassifier', createDocumentClassifier_languageCode - The language of the input documents. You can specify any of the following languages supported by Amazon Comprehend: German ("de"), English ("en"), Spanish ("es"), French ("fr"), Italian ("it"), or Portuguese ("pt"). All documents must be in the same language.

Request Lenses

createDocumentClassifier_versionName :: Lens' CreateDocumentClassifier (Maybe Text)

The version name given to the newly created classifier. Version names can have a maximum of 256 characters. Alphanumeric characters, hyphens (-) and underscores (_) are allowed. The version name must be unique among all models with the same classifier name in the account/AWS Region.

createDocumentClassifier_modelKmsKeyId :: Lens' CreateDocumentClassifier (Maybe Text)

ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt trained custom models. The ModelKmsKeyId can be either of the following formats:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

createDocumentClassifier_mode :: Lens' CreateDocumentClassifier (Maybe DocumentClassifierMode)

Indicates the mode in which the classifier will be trained. The classifier can be trained in multi-class mode, which identifies one and only one class for each document, or multi-label mode, which identifies one or more labels for each document. In multi-label mode, multiple labels for an individual document are separated by a delimiter. The default delimiter between labels is a pipe (|).

createDocumentClassifier_vpcConfig :: Lens' CreateDocumentClassifier (Maybe VpcConfig)

Configuration parameters for an optional private Virtual Private Cloud (VPC) containing the resources you are using for your custom classifier. For more information, see Amazon VPC.

createDocumentClassifier_volumeKmsKeyId :: Lens' CreateDocumentClassifier (Maybe Text)

ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt data on the storage volume attached to the ML compute instance(s) that process the analysis job. The VolumeKmsKeyId can be either of the following formats:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

createDocumentClassifier_outputDataConfig :: Lens' CreateDocumentClassifier (Maybe DocumentClassifierOutputDataConfig)

Enables the addition of output results configuration parameters for custom classifier jobs.

createDocumentClassifier_clientRequestToken :: Lens' CreateDocumentClassifier (Maybe Text)

A unique identifier for the request. If you don't set the client request token, Amazon Comprehend generates one.

createDocumentClassifier_tags :: Lens' CreateDocumentClassifier (Maybe [Tag])

Tags to be associated with the document classifier being created. A tag is a key-value pair that adds as a metadata to a resource used by Amazon Comprehend. For example, a tag with "Sales" as the key might be added to a resource to indicate its use by the sales department.

createDocumentClassifier_documentClassifierName :: Lens' CreateDocumentClassifier Text

The name of the document classifier.

createDocumentClassifier_dataAccessRoleArn :: Lens' CreateDocumentClassifier Text

The Amazon Resource Name (ARN) of the AWS Identity and Management (IAM) role that grants Amazon Comprehend read access to your input data.

createDocumentClassifier_inputDataConfig :: Lens' CreateDocumentClassifier DocumentClassifierInputDataConfig

Specifies the format and location of the input data for the job.

createDocumentClassifier_languageCode :: Lens' CreateDocumentClassifier LanguageCode

The language of the input documents. You can specify any of the following languages supported by Amazon Comprehend: German ("de"), English ("en"), Spanish ("es"), French ("fr"), Italian ("it"), or Portuguese ("pt"). All documents must be in the same language.

Destructuring the Response

data CreateDocumentClassifierResponse

See: newCreateDocumentClassifierResponse smart constructor.

Constructors

CreateDocumentClassifierResponse'

Fields

documentClassifierArn :: Maybe Text

The Amazon Resource Name (ARN) that identifies the document classifier.

httpStatus :: Int

The response's http status code.

newCreateDocumentClassifierResponse

Arguments

:: Int
$sel:httpStatus:CreateDocumentClassifierResponse'

-> CreateDocumentClassifierResponse

Create a value of CreateDocumentClassifierResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:documentClassifierArn:CreateDocumentClassifierResponse', createDocumentClassifierResponse_documentClassifierArn - The Amazon Resource Name (ARN) that identifies the document classifier.

$sel:httpStatus:CreateDocumentClassifierResponse', createDocumentClassifierResponse_httpStatus - The response's http status code.

Response Lenses

createDocumentClassifierResponse_documentClassifierArn :: Lens' CreateDocumentClassifierResponse (Maybe Text)

The Amazon Resource Name (ARN) that identifies the document classifier.

createDocumentClassifierResponse_httpStatus :: Lens' CreateDocumentClassifierResponse Int

The response's http status code.