Network.AWS.Comprehend.DeleteEntityRecognizer

Description

Deletes an entity recognizer.

Only those recognizers that are in terminated states (IN_ERROR, TRAINED) will be deleted. If an active inference job is using the model, a ResourceInUseException will be returned.

This is an asynchronous action that puts the recognizer into a DELETING state, and it is then removed by a background job. Once removed, the recognizer disappears from your account and is no longer available for use.

Creating a Request

data DeleteEntityRecognizer

See: newDeleteEntityRecognizer smart constructor.

Constructors

DeleteEntityRecognizer'

Fields

entityRecognizerArn :: Text

The Amazon Resource Name (ARN) that identifies the entity recognizer.

newDeleteEntityRecognizer

Arguments

:: Text
$sel:entityRecognizerArn:DeleteEntityRecognizer'

-> DeleteEntityRecognizer

Create a value of DeleteEntityRecognizer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:entityRecognizerArn:DeleteEntityRecognizer', deleteEntityRecognizer_entityRecognizerArn - The Amazon Resource Name (ARN) that identifies the entity recognizer.

Request Lenses

deleteEntityRecognizer_entityRecognizerArn :: Lens' DeleteEntityRecognizer Text

The Amazon Resource Name (ARN) that identifies the entity recognizer.

Destructuring the Response

data DeleteEntityRecognizerResponse

See: newDeleteEntityRecognizerResponse smart constructor.

Constructors

DeleteEntityRecognizerResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteEntityRecognizerResponse

Arguments

:: Int
$sel:httpStatus:DeleteEntityRecognizerResponse'

-> DeleteEntityRecognizerResponse

Create a value of DeleteEntityRecognizerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteEntityRecognizerResponse', deleteEntityRecognizerResponse_httpStatus - The response's http status code.

Response Lenses

deleteEntityRecognizerResponse_httpStatus :: Lens' DeleteEntityRecognizerResponse Int

The response's http status code.