Network.AWS.Comprehend.DescribeEntitiesDetectionJob

Description

Gets the properties associated with an entities detection job. Use this operation to get the status of a detection job.

Creating a Request

data DescribeEntitiesDetectionJob

See: newDescribeEntitiesDetectionJob smart constructor.

Constructors

DescribeEntitiesDetectionJob'

Fields

jobId :: Text

The identifier that Amazon Comprehend generated for the job. The operation returns this identifier in its response.

newDescribeEntitiesDetectionJob

Arguments

:: Text
$sel:jobId:DescribeEntitiesDetectionJob'

-> DescribeEntitiesDetectionJob

Create a value of DescribeEntitiesDetectionJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobId:DescribeEntitiesDetectionJob', describeEntitiesDetectionJob_jobId - The identifier that Amazon Comprehend generated for the job. The operation returns this identifier in its response.

Request Lenses

describeEntitiesDetectionJob_jobId :: Lens' DescribeEntitiesDetectionJob Text

The identifier that Amazon Comprehend generated for the job. The operation returns this identifier in its response.

Destructuring the Response

data DescribeEntitiesDetectionJobResponse

See: newDescribeEntitiesDetectionJobResponse smart constructor.

Constructors

DescribeEntitiesDetectionJobResponse'

Fields

entitiesDetectionJobProperties :: Maybe EntitiesDetectionJobProperties

An object that contains the properties associated with an entities detection job.

httpStatus :: Int

The response's http status code.

newDescribeEntitiesDetectionJobResponse

Arguments

:: Int
$sel:httpStatus:DescribeEntitiesDetectionJobResponse'

-> DescribeEntitiesDetectionJobResponse

Create a value of DescribeEntitiesDetectionJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:entitiesDetectionJobProperties:DescribeEntitiesDetectionJobResponse', describeEntitiesDetectionJobResponse_entitiesDetectionJobProperties - An object that contains the properties associated with an entities detection job.

$sel:httpStatus:DescribeEntitiesDetectionJobResponse', describeEntitiesDetectionJobResponse_httpStatus - The response's http status code.

Response Lenses

describeEntitiesDetectionJobResponse_entitiesDetectionJobProperties :: Lens' DescribeEntitiesDetectionJobResponse (Maybe EntitiesDetectionJobProperties)

An object that contains the properties associated with an entities detection job.

describeEntitiesDetectionJobResponse_httpStatus :: Lens' DescribeEntitiesDetectionJobResponse Int

The response's http status code.