Network.AWS.Comprehend.DetectDominantLanguage

Description

Determines the dominant language of the input text. For a list of languages that Amazon Comprehend can detect, see Amazon Comprehend Supported Languages.

Creating a Request

data DetectDominantLanguage

See: newDetectDominantLanguage smart constructor.

Constructors

DetectDominantLanguage'

Fields

text :: Sensitive Text

A UTF-8 text string. Each string should contain at least 20 characters and must contain fewer that 5,000 bytes of UTF-8 encoded characters.

newDetectDominantLanguage

Arguments

:: Text
$sel:text:DetectDominantLanguage'

-> DetectDominantLanguage

Create a value of DetectDominantLanguage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:text:DetectDominantLanguage', detectDominantLanguage_text - A UTF-8 text string. Each string should contain at least 20 characters and must contain fewer that 5,000 bytes of UTF-8 encoded characters.

Request Lenses

detectDominantLanguage_text :: Lens' DetectDominantLanguage Text

A UTF-8 text string. Each string should contain at least 20 characters and must contain fewer that 5,000 bytes of UTF-8 encoded characters.

Destructuring the Response

data DetectDominantLanguageResponse

See: newDetectDominantLanguageResponse smart constructor.

Constructors

DetectDominantLanguageResponse'

Fields

languages :: Maybe [DominantLanguage]

The languages that Amazon Comprehend detected in the input text. For each language, the response returns the RFC 5646 language code and the level of confidence that Amazon Comprehend has in the accuracy of its inference. For more information about RFC 5646, see Tags for Identifying Languages on the IETF Tools web site.

httpStatus :: Int

The response's http status code.

newDetectDominantLanguageResponse

Arguments

:: Int
$sel:httpStatus:DetectDominantLanguageResponse'

-> DetectDominantLanguageResponse

Create a value of DetectDominantLanguageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:languages:DetectDominantLanguageResponse', detectDominantLanguageResponse_languages - The languages that Amazon Comprehend detected in the input text. For each language, the response returns the RFC 5646 language code and the level of confidence that Amazon Comprehend has in the accuracy of its inference. For more information about RFC 5646, see Tags for Identifying Languages on the IETF Tools web site.

$sel:httpStatus:DetectDominantLanguageResponse', detectDominantLanguageResponse_httpStatus - The response's http status code.

Response Lenses

detectDominantLanguageResponse_languages :: Lens' DetectDominantLanguageResponse (Maybe [DominantLanguage])

The languages that Amazon Comprehend detected in the input text. For each language, the response returns the RFC 5646 language code and the level of confidence that Amazon Comprehend has in the accuracy of its inference. For more information about RFC 5646, see Tags for Identifying Languages on the IETF Tools web site.

detectDominantLanguageResponse_httpStatus :: Lens' DetectDominantLanguageResponse Int

The response's http status code.