Network.AWS.Comprehend.DetectKeyPhrases

Description

Detects the key noun phrases found in the text.

Creating a Request

data DetectKeyPhrases

See: newDetectKeyPhrases smart constructor.

Constructors

DetectKeyPhrases'

Fields

text :: Sensitive Text

A UTF-8 text string. Each string must contain fewer that 5,000 bytes of UTF-8 encoded characters.

languageCode :: LanguageCode

The language of the input documents. You can specify any of the primary languages supported by Amazon Comprehend. All documents must be in the same language.

newDetectKeyPhrases

Arguments

:: Text
$sel:text:DetectKeyPhrases'

-> LanguageCode
$sel:languageCode:DetectKeyPhrases'

-> DetectKeyPhrases

Create a value of DetectKeyPhrases with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:text:DetectKeyPhrases', detectKeyPhrases_text - A UTF-8 text string. Each string must contain fewer that 5,000 bytes of UTF-8 encoded characters.

$sel:languageCode:DetectKeyPhrases', detectKeyPhrases_languageCode - The language of the input documents. You can specify any of the primary languages supported by Amazon Comprehend. All documents must be in the same language.

Request Lenses

detectKeyPhrases_text :: Lens' DetectKeyPhrases Text

A UTF-8 text string. Each string must contain fewer that 5,000 bytes of UTF-8 encoded characters.

detectKeyPhrases_languageCode :: Lens' DetectKeyPhrases LanguageCode

The language of the input documents. You can specify any of the primary languages supported by Amazon Comprehend. All documents must be in the same language.

Destructuring the Response

data DetectKeyPhrasesResponse

See: newDetectKeyPhrasesResponse smart constructor.

Constructors

DetectKeyPhrasesResponse'

Fields

keyPhrases :: Maybe [KeyPhrase]

A collection of key phrases that Amazon Comprehend identified in the input text. For each key phrase, the response provides the text of the key phrase, where the key phrase begins and ends, and the level of confidence that Amazon Comprehend has in the accuracy of the detection.

httpStatus :: Int

The response's http status code.

newDetectKeyPhrasesResponse

Arguments

:: Int
$sel:httpStatus:DetectKeyPhrasesResponse'

-> DetectKeyPhrasesResponse

Create a value of DetectKeyPhrasesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:keyPhrases:DetectKeyPhrasesResponse', detectKeyPhrasesResponse_keyPhrases - A collection of key phrases that Amazon Comprehend identified in the input text. For each key phrase, the response provides the text of the key phrase, where the key phrase begins and ends, and the level of confidence that Amazon Comprehend has in the accuracy of the detection.

$sel:httpStatus:DetectKeyPhrasesResponse', detectKeyPhrasesResponse_httpStatus - The response's http status code.

Response Lenses

detectKeyPhrasesResponse_keyPhrases :: Lens' DetectKeyPhrasesResponse (Maybe [KeyPhrase])

A collection of key phrases that Amazon Comprehend identified in the input text. For each key phrase, the response provides the text of the key phrase, where the key phrase begins and ends, and the level of confidence that Amazon Comprehend has in the accuracy of the detection.

detectKeyPhrasesResponse_httpStatus :: Lens' DetectKeyPhrasesResponse Int

The response's http status code.