Network.AWS.Comprehend.DetectSentiment

Description

Inspects text and returns an inference of the prevailing sentiment (POSITIVE, NEUTRAL, MIXED, or NEGATIVE).

Creating a Request

data DetectSentiment

See: newDetectSentiment smart constructor.

Constructors

DetectSentiment'

Fields

text :: Sensitive Text

A UTF-8 text string. Each string must contain fewer that 5,000 bytes of UTF-8 encoded characters.

languageCode :: LanguageCode

The language of the input documents. You can specify any of the primary languages supported by Amazon Comprehend. All documents must be in the same language.

newDetectSentiment

Arguments

:: Text
$sel:text:DetectSentiment'

-> LanguageCode
$sel:languageCode:DetectSentiment'

-> DetectSentiment

Create a value of DetectSentiment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:text:DetectSentiment', detectSentiment_text - A UTF-8 text string. Each string must contain fewer that 5,000 bytes of UTF-8 encoded characters.

$sel:languageCode:DetectSentiment', detectSentiment_languageCode - The language of the input documents. You can specify any of the primary languages supported by Amazon Comprehend. All documents must be in the same language.

Request Lenses

detectSentiment_text :: Lens' DetectSentiment Text

A UTF-8 text string. Each string must contain fewer that 5,000 bytes of UTF-8 encoded characters.

detectSentiment_languageCode :: Lens' DetectSentiment LanguageCode

The language of the input documents. You can specify any of the primary languages supported by Amazon Comprehend. All documents must be in the same language.

Destructuring the Response

data DetectSentimentResponse

See: newDetectSentimentResponse smart constructor.

Constructors

DetectSentimentResponse'

Fields

sentiment :: Maybe SentimentType

The inferred sentiment that Amazon Comprehend has the highest level of confidence in.

sentimentScore :: Maybe SentimentScore

An object that lists the sentiments, and their corresponding confidence levels.

httpStatus :: Int

The response's http status code.

newDetectSentimentResponse

Arguments

:: Int
$sel:httpStatus:DetectSentimentResponse'

-> DetectSentimentResponse

Create a value of DetectSentimentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sentiment:DetectSentimentResponse', detectSentimentResponse_sentiment - The inferred sentiment that Amazon Comprehend has the highest level of confidence in.

$sel:sentimentScore:DetectSentimentResponse', detectSentimentResponse_sentimentScore - An object that lists the sentiments, and their corresponding confidence levels.

$sel:httpStatus:DetectSentimentResponse', detectSentimentResponse_httpStatus - The response's http status code.

Response Lenses

detectSentimentResponse_sentiment :: Lens' DetectSentimentResponse (Maybe SentimentType)

The inferred sentiment that Amazon Comprehend has the highest level of confidence in.

detectSentimentResponse_sentimentScore :: Lens' DetectSentimentResponse (Maybe SentimentScore)

An object that lists the sentiments, and their corresponding confidence levels.

detectSentimentResponse_httpStatus :: Lens' DetectSentimentResponse Int

The response's http status code.