Network.AWS.Comprehend.ListEntitiesDetectionJobs

Description

Gets a list of the entity detection jobs that you have submitted.

This operation returns paginated results.

Creating a Request

data ListEntitiesDetectionJobs

See: newListEntitiesDetectionJobs smart constructor.

Constructors

ListEntitiesDetectionJobs'

Fields

nextToken :: Maybe Text

Identifies the next page of results to return.

maxResults :: Maybe Natural

The maximum number of results to return in each page. The default is 100.

filter' :: Maybe EntitiesDetectionJobFilter

Filters the jobs that are returned. You can filter jobs on their name, status, or the date and time that they were submitted. You can only set one filter at a time.

newListEntitiesDetectionJobs :: ListEntitiesDetectionJobs

Create a value of ListEntitiesDetectionJobs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListEntitiesDetectionJobs', listEntitiesDetectionJobs_nextToken - Identifies the next page of results to return.

$sel:maxResults:ListEntitiesDetectionJobs', listEntitiesDetectionJobs_maxResults - The maximum number of results to return in each page. The default is 100.

$sel:filter':ListEntitiesDetectionJobs', listEntitiesDetectionJobs_filter - Filters the jobs that are returned. You can filter jobs on their name, status, or the date and time that they were submitted. You can only set one filter at a time.

Request Lenses

listEntitiesDetectionJobs_nextToken :: Lens' ListEntitiesDetectionJobs (Maybe Text)

Identifies the next page of results to return.

listEntitiesDetectionJobs_maxResults :: Lens' ListEntitiesDetectionJobs (Maybe Natural)

The maximum number of results to return in each page. The default is 100.

listEntitiesDetectionJobs_filter :: Lens' ListEntitiesDetectionJobs (Maybe EntitiesDetectionJobFilter)

Filters the jobs that are returned. You can filter jobs on their name, status, or the date and time that they were submitted. You can only set one filter at a time.

Destructuring the Response

data ListEntitiesDetectionJobsResponse

See: newListEntitiesDetectionJobsResponse smart constructor.

Constructors

ListEntitiesDetectionJobsResponse'

Fields

nextToken :: Maybe Text

Identifies the next page of results to return.

entitiesDetectionJobPropertiesList :: Maybe [EntitiesDetectionJobProperties]

A list containing the properties of each job that is returned.

httpStatus :: Int

The response's http status code.

newListEntitiesDetectionJobsResponse

Arguments

:: Int
$sel:httpStatus:ListEntitiesDetectionJobsResponse'

-> ListEntitiesDetectionJobsResponse

Create a value of ListEntitiesDetectionJobsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListEntitiesDetectionJobs', listEntitiesDetectionJobsResponse_nextToken - Identifies the next page of results to return.

$sel:entitiesDetectionJobPropertiesList:ListEntitiesDetectionJobsResponse', listEntitiesDetectionJobsResponse_entitiesDetectionJobPropertiesList - A list containing the properties of each job that is returned.

$sel:httpStatus:ListEntitiesDetectionJobsResponse', listEntitiesDetectionJobsResponse_httpStatus - The response's http status code.

Response Lenses

listEntitiesDetectionJobsResponse_nextToken :: Lens' ListEntitiesDetectionJobsResponse (Maybe Text)

Identifies the next page of results to return.

listEntitiesDetectionJobsResponse_entitiesDetectionJobPropertiesList :: Lens' ListEntitiesDetectionJobsResponse (Maybe [EntitiesDetectionJobProperties])

A list containing the properties of each job that is returned.

listEntitiesDetectionJobsResponse_httpStatus :: Lens' ListEntitiesDetectionJobsResponse Int

The response's http status code.