Network.AWS.Comprehend.ListEntityRecognizerSummaries

Description

Gets a list of summaries for the entity recognizers that you have created.

Creating a Request

data ListEntityRecognizerSummaries

See: newListEntityRecognizerSummaries smart constructor.

Constructors

ListEntityRecognizerSummaries'

Fields

nextToken :: Maybe Text

Identifies the next page of results to return.

maxResults :: Maybe Natural

The maximum number of results to return on each page. The default is 100.

newListEntityRecognizerSummaries :: ListEntityRecognizerSummaries

Create a value of ListEntityRecognizerSummaries with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListEntityRecognizerSummaries', listEntityRecognizerSummaries_nextToken - Identifies the next page of results to return.

$sel:maxResults:ListEntityRecognizerSummaries', listEntityRecognizerSummaries_maxResults - The maximum number of results to return on each page. The default is 100.

Request Lenses

listEntityRecognizerSummaries_nextToken :: Lens' ListEntityRecognizerSummaries (Maybe Text)

Identifies the next page of results to return.

listEntityRecognizerSummaries_maxResults :: Lens' ListEntityRecognizerSummaries (Maybe Natural)

The maximum number of results to return on each page. The default is 100.

Destructuring the Response

data ListEntityRecognizerSummariesResponse

See: newListEntityRecognizerSummariesResponse smart constructor.

Constructors

ListEntityRecognizerSummariesResponse'

Fields

nextToken :: Maybe Text

The list entity recognizer summaries.

entityRecognizerSummariesList :: Maybe [EntityRecognizerSummary]

The list entity recognizer summaries.

httpStatus :: Int

The response's http status code.

newListEntityRecognizerSummariesResponse

Arguments

:: Int
$sel:httpStatus:ListEntityRecognizerSummariesResponse'

-> ListEntityRecognizerSummariesResponse

Create a value of ListEntityRecognizerSummariesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListEntityRecognizerSummaries', listEntityRecognizerSummariesResponse_nextToken - The list entity recognizer summaries.

$sel:entityRecognizerSummariesList:ListEntityRecognizerSummariesResponse', listEntityRecognizerSummariesResponse_entityRecognizerSummariesList - The list entity recognizer summaries.

$sel:httpStatus:ListEntityRecognizerSummariesResponse', listEntityRecognizerSummariesResponse_httpStatus - The response's http status code.

Response Lenses

listEntityRecognizerSummariesResponse_nextToken :: Lens' ListEntityRecognizerSummariesResponse (Maybe Text)

The list entity recognizer summaries.

listEntityRecognizerSummariesResponse_entityRecognizerSummariesList :: Lens' ListEntityRecognizerSummariesResponse (Maybe [EntityRecognizerSummary])

The list entity recognizer summaries.

listEntityRecognizerSummariesResponse_httpStatus :: Lens' ListEntityRecognizerSummariesResponse Int

The response's http status code.