Network.AWS.Comprehend.ListSentimentDetectionJobs

Description

Gets a list of sentiment detection jobs that you have submitted.

This operation returns paginated results.

Creating a Request

data ListSentimentDetectionJobs

See: newListSentimentDetectionJobs smart constructor.

Constructors

ListSentimentDetectionJobs'

Fields

nextToken :: Maybe Text

Identifies the next page of results to return.

maxResults :: Maybe Natural

The maximum number of results to return in each page. The default is 100.

filter' :: Maybe SentimentDetectionJobFilter

Filters the jobs that are returned. You can filter jobs on their name, status, or the date and time that they were submitted. You can only set one filter at a time.

newListSentimentDetectionJobs :: ListSentimentDetectionJobs

Create a value of ListSentimentDetectionJobs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListSentimentDetectionJobs', listSentimentDetectionJobs_nextToken - Identifies the next page of results to return.

$sel:maxResults:ListSentimentDetectionJobs', listSentimentDetectionJobs_maxResults - The maximum number of results to return in each page. The default is 100.

$sel:filter':ListSentimentDetectionJobs', listSentimentDetectionJobs_filter - Filters the jobs that are returned. You can filter jobs on their name, status, or the date and time that they were submitted. You can only set one filter at a time.

Request Lenses

listSentimentDetectionJobs_nextToken :: Lens' ListSentimentDetectionJobs (Maybe Text)

Identifies the next page of results to return.

listSentimentDetectionJobs_maxResults :: Lens' ListSentimentDetectionJobs (Maybe Natural)

The maximum number of results to return in each page. The default is 100.

listSentimentDetectionJobs_filter :: Lens' ListSentimentDetectionJobs (Maybe SentimentDetectionJobFilter)

Filters the jobs that are returned. You can filter jobs on their name, status, or the date and time that they were submitted. You can only set one filter at a time.

Destructuring the Response

data ListSentimentDetectionJobsResponse

See: newListSentimentDetectionJobsResponse smart constructor.

Constructors

ListSentimentDetectionJobsResponse'

Fields

sentimentDetectionJobPropertiesList :: Maybe [SentimentDetectionJobProperties]

A list containing the properties of each job that is returned.

nextToken :: Maybe Text

Identifies the next page of results to return.

httpStatus :: Int

The response's http status code.

newListSentimentDetectionJobsResponse

Arguments

:: Int
$sel:httpStatus:ListSentimentDetectionJobsResponse'

-> ListSentimentDetectionJobsResponse

Create a value of ListSentimentDetectionJobsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sentimentDetectionJobPropertiesList:ListSentimentDetectionJobsResponse', listSentimentDetectionJobsResponse_sentimentDetectionJobPropertiesList - A list containing the properties of each job that is returned.

$sel:nextToken:ListSentimentDetectionJobs', listSentimentDetectionJobsResponse_nextToken - Identifies the next page of results to return.

$sel:httpStatus:ListSentimentDetectionJobsResponse', listSentimentDetectionJobsResponse_httpStatus - The response's http status code.

Response Lenses

listSentimentDetectionJobsResponse_sentimentDetectionJobPropertiesList :: Lens' ListSentimentDetectionJobsResponse (Maybe [SentimentDetectionJobProperties])

A list containing the properties of each job that is returned.

listSentimentDetectionJobsResponse_nextToken :: Lens' ListSentimentDetectionJobsResponse (Maybe Text)

Identifies the next page of results to return.

listSentimentDetectionJobsResponse_httpStatus :: Lens' ListSentimentDetectionJobsResponse Int

The response's http status code.