Network.AWS.Comprehend.StartEntitiesDetectionJob

Description

Starts an asynchronous entity detection job for a collection of documents. Use the operation to track the status of a job.

This API can be used for either standard entity detection or custom entity recognition. In order to be used for custom entity recognition, the optional EntityRecognizerArn must be used in order to provide access to the recognizer being used to detect the custom entity.

Creating a Request

data StartEntitiesDetectionJob

See: newStartEntitiesDetectionJob smart constructor.

Constructors

StartEntitiesDetectionJob'

Fields

vpcConfig :: Maybe VpcConfig

Configuration parameters for an optional private Virtual Private Cloud (VPC) containing the resources you are using for your entity detection job. For more information, see Amazon VPC.

volumeKmsKeyId :: Maybe Text

ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt data on the storage volume attached to the ML compute instance(s) that process the analysis job. The VolumeKmsKeyId can be either of the following formats:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

tags :: Maybe [Tag]

Tags to be associated with the entities detection job. A tag is a key-value pair that adds metadata to a resource used by Amazon Comprehend. For example, a tag with "Sales" as the key might be added to a resource to indicate its use by the sales department.

clientRequestToken :: Maybe Text

A unique identifier for the request. If you don't set the client request token, Amazon Comprehend generates one.

entityRecognizerArn :: Maybe Text

The Amazon Resource Name (ARN) that identifies the specific entity recognizer to be used by the StartEntitiesDetectionJob. This ARN is optional and is only used for a custom entity recognition job.

jobName :: Maybe Text

The identifier of the job.

inputDataConfig :: InputDataConfig

Specifies the format and location of the input data for the job.

outputDataConfig :: OutputDataConfig

Specifies where to send the output files.

dataAccessRoleArn :: Text

The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role that grants Amazon Comprehend read access to your input data. For more information, see https://docs.aws.amazon.com/comprehend/latest/dg/access-control-managing-permissions.html#auth-role-permissions.

languageCode :: LanguageCode

The language of the input documents. All documents must be in the same language. You can specify any of the languages supported by Amazon Comprehend. If custom entities recognition is used, this parameter is ignored and the language used for training the model is used instead.

newStartEntitiesDetectionJob

Arguments

:: InputDataConfig
$sel:inputDataConfig:StartEntitiesDetectionJob'

-> OutputDataConfig
$sel:outputDataConfig:StartEntitiesDetectionJob'

-> Text
$sel:dataAccessRoleArn:StartEntitiesDetectionJob'

-> LanguageCode
$sel:languageCode:StartEntitiesDetectionJob'

-> StartEntitiesDetectionJob

Create a value of StartEntitiesDetectionJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpcConfig:StartEntitiesDetectionJob', startEntitiesDetectionJob_vpcConfig - Configuration parameters for an optional private Virtual Private Cloud (VPC) containing the resources you are using for your entity detection job. For more information, see Amazon VPC.

$sel:volumeKmsKeyId:StartEntitiesDetectionJob', startEntitiesDetectionJob_volumeKmsKeyId - ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt data on the storage volume attached to the ML compute instance(s) that process the analysis job. The VolumeKmsKeyId can be either of the following formats:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

$sel:tags:StartEntitiesDetectionJob', startEntitiesDetectionJob_tags - Tags to be associated with the entities detection job. A tag is a key-value pair that adds metadata to a resource used by Amazon Comprehend. For example, a tag with "Sales" as the key might be added to a resource to indicate its use by the sales department.

$sel:clientRequestToken:StartEntitiesDetectionJob', startEntitiesDetectionJob_clientRequestToken - A unique identifier for the request. If you don't set the client request token, Amazon Comprehend generates one.

$sel:entityRecognizerArn:StartEntitiesDetectionJob', startEntitiesDetectionJob_entityRecognizerArn - The Amazon Resource Name (ARN) that identifies the specific entity recognizer to be used by the StartEntitiesDetectionJob. This ARN is optional and is only used for a custom entity recognition job.

$sel:jobName:StartEntitiesDetectionJob', startEntitiesDetectionJob_jobName - The identifier of the job.

$sel:inputDataConfig:StartEntitiesDetectionJob', startEntitiesDetectionJob_inputDataConfig - Specifies the format and location of the input data for the job.

$sel:outputDataConfig:StartEntitiesDetectionJob', startEntitiesDetectionJob_outputDataConfig - Specifies where to send the output files.

$sel:dataAccessRoleArn:StartEntitiesDetectionJob', startEntitiesDetectionJob_dataAccessRoleArn - The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role that grants Amazon Comprehend read access to your input data. For more information, see https://docs.aws.amazon.com/comprehend/latest/dg/access-control-managing-permissions.html#auth-role-permissions.

$sel:languageCode:StartEntitiesDetectionJob', startEntitiesDetectionJob_languageCode - The language of the input documents. All documents must be in the same language. You can specify any of the languages supported by Amazon Comprehend. If custom entities recognition is used, this parameter is ignored and the language used for training the model is used instead.

Request Lenses

startEntitiesDetectionJob_vpcConfig :: Lens' StartEntitiesDetectionJob (Maybe VpcConfig)

Configuration parameters for an optional private Virtual Private Cloud (VPC) containing the resources you are using for your entity detection job. For more information, see Amazon VPC.

startEntitiesDetectionJob_volumeKmsKeyId :: Lens' StartEntitiesDetectionJob (Maybe Text)

ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt data on the storage volume attached to the ML compute instance(s) that process the analysis job. The VolumeKmsKeyId can be either of the following formats:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

startEntitiesDetectionJob_tags :: Lens' StartEntitiesDetectionJob (Maybe [Tag])

Tags to be associated with the entities detection job. A tag is a key-value pair that adds metadata to a resource used by Amazon Comprehend. For example, a tag with "Sales" as the key might be added to a resource to indicate its use by the sales department.

startEntitiesDetectionJob_clientRequestToken :: Lens' StartEntitiesDetectionJob (Maybe Text)

A unique identifier for the request. If you don't set the client request token, Amazon Comprehend generates one.

startEntitiesDetectionJob_entityRecognizerArn :: Lens' StartEntitiesDetectionJob (Maybe Text)

The Amazon Resource Name (ARN) that identifies the specific entity recognizer to be used by the StartEntitiesDetectionJob. This ARN is optional and is only used for a custom entity recognition job.

startEntitiesDetectionJob_jobName :: Lens' StartEntitiesDetectionJob (Maybe Text)

The identifier of the job.

startEntitiesDetectionJob_inputDataConfig :: Lens' StartEntitiesDetectionJob InputDataConfig

Specifies the format and location of the input data for the job.

startEntitiesDetectionJob_outputDataConfig :: Lens' StartEntitiesDetectionJob OutputDataConfig

Specifies where to send the output files.

startEntitiesDetectionJob_dataAccessRoleArn :: Lens' StartEntitiesDetectionJob Text

The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role that grants Amazon Comprehend read access to your input data. For more information, see https://docs.aws.amazon.com/comprehend/latest/dg/access-control-managing-permissions.html#auth-role-permissions.

startEntitiesDetectionJob_languageCode :: Lens' StartEntitiesDetectionJob LanguageCode

The language of the input documents. All documents must be in the same language. You can specify any of the languages supported by Amazon Comprehend. If custom entities recognition is used, this parameter is ignored and the language used for training the model is used instead.

Destructuring the Response

data StartEntitiesDetectionJobResponse

See: newStartEntitiesDetectionJobResponse smart constructor.

Constructors

StartEntitiesDetectionJobResponse'

Fields

jobStatus :: Maybe JobStatus

The status of the job.

SUBMITTED - The job has been received and is queued for processing.

IN_PROGRESS - Amazon Comprehend is processing the job.

COMPLETED - The job was successfully completed and the output is available.

FAILED - The job did not complete. To get details, use the operation.

STOP_REQUESTED - Amazon Comprehend has received a stop request for the job and is processing the request.

STOPPED - The job was successfully stopped without completing.

jobArn :: Maybe Text

The Amazon Resource Name (ARN) of the entities detection job. It is a unique, fully qualified identifier for the job. It includes the AWS account, Region, and the job ID. The format of the ARN is as follows:

arn:<partition>:comprehend:<region>:<account-id>:entities-detection-job/<job-id>

The following is an example job ARN:

arn:aws:comprehend:us-west-2:111122223333:entities-detection-job/1234abcd12ab34cd56ef1234567890ab

jobId :: Maybe Text

The identifier generated for the job. To get the status of job, use this identifier with the operation.

httpStatus :: Int

The response's http status code.

newStartEntitiesDetectionJobResponse

Arguments

:: Int
$sel:httpStatus:StartEntitiesDetectionJobResponse'

-> StartEntitiesDetectionJobResponse

Create a value of StartEntitiesDetectionJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobStatus:StartEntitiesDetectionJobResponse', startEntitiesDetectionJobResponse_jobStatus - The status of the job.

SUBMITTED - The job has been received and is queued for processing.

IN_PROGRESS - Amazon Comprehend is processing the job.

COMPLETED - The job was successfully completed and the output is available.

FAILED - The job did not complete. To get details, use the operation.

STOP_REQUESTED - Amazon Comprehend has received a stop request for the job and is processing the request.

STOPPED - The job was successfully stopped without completing.

$sel:jobArn:StartEntitiesDetectionJobResponse', startEntitiesDetectionJobResponse_jobArn - The Amazon Resource Name (ARN) of the entities detection job. It is a unique, fully qualified identifier for the job. It includes the AWS account, Region, and the job ID. The format of the ARN is as follows:

arn:<partition>:comprehend:<region>:<account-id>:entities-detection-job/<job-id>

The following is an example job ARN:

arn:aws:comprehend:us-west-2:111122223333:entities-detection-job/1234abcd12ab34cd56ef1234567890ab

$sel:jobId:StartEntitiesDetectionJobResponse', startEntitiesDetectionJobResponse_jobId - The identifier generated for the job. To get the status of job, use this identifier with the operation.

$sel:httpStatus:StartEntitiesDetectionJobResponse', startEntitiesDetectionJobResponse_httpStatus - The response's http status code.

Response Lenses

startEntitiesDetectionJobResponse_jobStatus :: Lens' StartEntitiesDetectionJobResponse (Maybe JobStatus)

The status of the job.

SUBMITTED - The job has been received and is queued for processing.

IN_PROGRESS - Amazon Comprehend is processing the job.

COMPLETED - The job was successfully completed and the output is available.

FAILED - The job did not complete. To get details, use the operation.

STOP_REQUESTED - Amazon Comprehend has received a stop request for the job and is processing the request.

STOPPED - The job was successfully stopped without completing.

startEntitiesDetectionJobResponse_jobArn :: Lens' StartEntitiesDetectionJobResponse (Maybe Text)

The Amazon Resource Name (ARN) of the entities detection job. It is a unique, fully qualified identifier for the job. It includes the AWS account, Region, and the job ID. The format of the ARN is as follows:

arn:<partition>:comprehend:<region>:<account-id>:entities-detection-job/<job-id>

The following is an example job ARN:

arn:aws:comprehend:us-west-2:111122223333:entities-detection-job/1234abcd12ab34cd56ef1234567890ab

startEntitiesDetectionJobResponse_jobId :: Lens' StartEntitiesDetectionJobResponse (Maybe Text)

The identifier generated for the job. To get the status of job, use this identifier with the operation.

startEntitiesDetectionJobResponse_httpStatus :: Lens' StartEntitiesDetectionJobResponse Int

The response's http status code.