Network.AWS.Comprehend.StartSentimentDetectionJob

Description

Starts an asynchronous sentiment detection job for a collection of documents. use the operation to track the status of a job.

Creating a Request

data StartSentimentDetectionJob

See: newStartSentimentDetectionJob smart constructor.

Constructors

StartSentimentDetectionJob'

Fields

vpcConfig :: Maybe VpcConfig

Configuration parameters for an optional private Virtual Private Cloud (VPC) containing the resources you are using for your sentiment detection job. For more information, see Amazon VPC.

volumeKmsKeyId :: Maybe Text

ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt data on the storage volume attached to the ML compute instance(s) that process the analysis job. The VolumeKmsKeyId can be either of the following formats:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

tags :: Maybe [Tag]

Tags to be associated with the sentiment detection job. A tag is a key-value pair that adds metadata to a resource used by Amazon Comprehend. For example, a tag with "Sales" as the key might be added to a resource to indicate its use by the sales department.

clientRequestToken :: Maybe Text

A unique identifier for the request. If you don't set the client request token, Amazon Comprehend generates one.

jobName :: Maybe Text

The identifier of the job.

inputDataConfig :: InputDataConfig

Specifies the format and location of the input data for the job.

outputDataConfig :: OutputDataConfig

Specifies where to send the output files.

dataAccessRoleArn :: Text

The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role that grants Amazon Comprehend read access to your input data. For more information, see https://docs.aws.amazon.com/comprehend/latest/dg/access-control-managing-permissions.html#auth-role-permissions.

languageCode :: LanguageCode

The language of the input documents. You can specify any of the primary languages supported by Amazon Comprehend. All documents must be in the same language.

newStartSentimentDetectionJob

Arguments

:: InputDataConfig
$sel:inputDataConfig:StartSentimentDetectionJob'

-> OutputDataConfig
$sel:outputDataConfig:StartSentimentDetectionJob'

-> Text
$sel:dataAccessRoleArn:StartSentimentDetectionJob'

-> LanguageCode
$sel:languageCode:StartSentimentDetectionJob'

-> StartSentimentDetectionJob

Create a value of StartSentimentDetectionJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpcConfig:StartSentimentDetectionJob', startSentimentDetectionJob_vpcConfig - Configuration parameters for an optional private Virtual Private Cloud (VPC) containing the resources you are using for your sentiment detection job. For more information, see Amazon VPC.

$sel:volumeKmsKeyId:StartSentimentDetectionJob', startSentimentDetectionJob_volumeKmsKeyId - ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt data on the storage volume attached to the ML compute instance(s) that process the analysis job. The VolumeKmsKeyId can be either of the following formats:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

$sel:tags:StartSentimentDetectionJob', startSentimentDetectionJob_tags - Tags to be associated with the sentiment detection job. A tag is a key-value pair that adds metadata to a resource used by Amazon Comprehend. For example, a tag with "Sales" as the key might be added to a resource to indicate its use by the sales department.

$sel:clientRequestToken:StartSentimentDetectionJob', startSentimentDetectionJob_clientRequestToken - A unique identifier for the request. If you don't set the client request token, Amazon Comprehend generates one.

$sel:jobName:StartSentimentDetectionJob', startSentimentDetectionJob_jobName - The identifier of the job.

$sel:inputDataConfig:StartSentimentDetectionJob', startSentimentDetectionJob_inputDataConfig - Specifies the format and location of the input data for the job.

$sel:outputDataConfig:StartSentimentDetectionJob', startSentimentDetectionJob_outputDataConfig - Specifies where to send the output files.

$sel:dataAccessRoleArn:StartSentimentDetectionJob', startSentimentDetectionJob_dataAccessRoleArn - The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role that grants Amazon Comprehend read access to your input data. For more information, see https://docs.aws.amazon.com/comprehend/latest/dg/access-control-managing-permissions.html#auth-role-permissions.

$sel:languageCode:StartSentimentDetectionJob', startSentimentDetectionJob_languageCode - The language of the input documents. You can specify any of the primary languages supported by Amazon Comprehend. All documents must be in the same language.

Request Lenses

startSentimentDetectionJob_vpcConfig :: Lens' StartSentimentDetectionJob (Maybe VpcConfig)

Configuration parameters for an optional private Virtual Private Cloud (VPC) containing the resources you are using for your sentiment detection job. For more information, see Amazon VPC.

startSentimentDetectionJob_volumeKmsKeyId :: Lens' StartSentimentDetectionJob (Maybe Text)

ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt data on the storage volume attached to the ML compute instance(s) that process the analysis job. The VolumeKmsKeyId can be either of the following formats:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

startSentimentDetectionJob_tags :: Lens' StartSentimentDetectionJob (Maybe [Tag])

Tags to be associated with the sentiment detection job. A tag is a key-value pair that adds metadata to a resource used by Amazon Comprehend. For example, a tag with "Sales" as the key might be added to a resource to indicate its use by the sales department.

startSentimentDetectionJob_clientRequestToken :: Lens' StartSentimentDetectionJob (Maybe Text)

A unique identifier for the request. If you don't set the client request token, Amazon Comprehend generates one.

startSentimentDetectionJob_jobName :: Lens' StartSentimentDetectionJob (Maybe Text)

The identifier of the job.

startSentimentDetectionJob_inputDataConfig :: Lens' StartSentimentDetectionJob InputDataConfig

Specifies the format and location of the input data for the job.

startSentimentDetectionJob_outputDataConfig :: Lens' StartSentimentDetectionJob OutputDataConfig

Specifies where to send the output files.

startSentimentDetectionJob_dataAccessRoleArn :: Lens' StartSentimentDetectionJob Text

The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role that grants Amazon Comprehend read access to your input data. For more information, see https://docs.aws.amazon.com/comprehend/latest/dg/access-control-managing-permissions.html#auth-role-permissions.

startSentimentDetectionJob_languageCode :: Lens' StartSentimentDetectionJob LanguageCode

The language of the input documents. You can specify any of the primary languages supported by Amazon Comprehend. All documents must be in the same language.

Destructuring the Response

data StartSentimentDetectionJobResponse

See: newStartSentimentDetectionJobResponse smart constructor.

Constructors

StartSentimentDetectionJobResponse'

Fields

jobStatus :: Maybe JobStatus

The status of the job.

SUBMITTED - The job has been received and is queued for processing.

IN_PROGRESS - Amazon Comprehend is processing the job.

COMPLETED - The job was successfully completed and the output is available.

FAILED - The job did not complete. To get details, use the operation.

jobArn :: Maybe Text

The Amazon Resource Name (ARN) of the sentiment detection job. It is a unique, fully qualified identifier for the job. It includes the AWS account, Region, and the job ID. The format of the ARN is as follows:

arn:<partition>:comprehend:<region>:<account-id>:sentiment-detection-job/<job-id>

The following is an example job ARN:

arn:aws:comprehend:us-west-2:111122223333:sentiment-detection-job/1234abcd12ab34cd56ef1234567890ab

jobId :: Maybe Text

The identifier generated for the job. To get the status of a job, use this identifier with the operation.

httpStatus :: Int

The response's http status code.

newStartSentimentDetectionJobResponse

Arguments

:: Int
$sel:httpStatus:StartSentimentDetectionJobResponse'

-> StartSentimentDetectionJobResponse

Create a value of StartSentimentDetectionJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobStatus:StartSentimentDetectionJobResponse', startSentimentDetectionJobResponse_jobStatus - The status of the job.

SUBMITTED - The job has been received and is queued for processing.

IN_PROGRESS - Amazon Comprehend is processing the job.

COMPLETED - The job was successfully completed and the output is available.

FAILED - The job did not complete. To get details, use the operation.

$sel:jobArn:StartSentimentDetectionJobResponse', startSentimentDetectionJobResponse_jobArn - The Amazon Resource Name (ARN) of the sentiment detection job. It is a unique, fully qualified identifier for the job. It includes the AWS account, Region, and the job ID. The format of the ARN is as follows:

arn:<partition>:comprehend:<region>:<account-id>:sentiment-detection-job/<job-id>

The following is an example job ARN:

arn:aws:comprehend:us-west-2:111122223333:sentiment-detection-job/1234abcd12ab34cd56ef1234567890ab

$sel:jobId:StartSentimentDetectionJobResponse', startSentimentDetectionJobResponse_jobId - The identifier generated for the job. To get the status of a job, use this identifier with the operation.

$sel:httpStatus:StartSentimentDetectionJobResponse', startSentimentDetectionJobResponse_httpStatus - The response's http status code.

Response Lenses

startSentimentDetectionJobResponse_jobStatus :: Lens' StartSentimentDetectionJobResponse (Maybe JobStatus)

The status of the job.

SUBMITTED - The job has been received and is queued for processing.

IN_PROGRESS - Amazon Comprehend is processing the job.

COMPLETED - The job was successfully completed and the output is available.

FAILED - The job did not complete. To get details, use the operation.

startSentimentDetectionJobResponse_jobArn :: Lens' StartSentimentDetectionJobResponse (Maybe Text)

The Amazon Resource Name (ARN) of the sentiment detection job. It is a unique, fully qualified identifier for the job. It includes the AWS account, Region, and the job ID. The format of the ARN is as follows:

arn:<partition>:comprehend:<region>:<account-id>:sentiment-detection-job/<job-id>

The following is an example job ARN:

arn:aws:comprehend:us-west-2:111122223333:sentiment-detection-job/1234abcd12ab34cd56ef1234567890ab

startSentimentDetectionJobResponse_jobId :: Lens' StartSentimentDetectionJobResponse (Maybe Text)

The identifier generated for the job. To get the status of a job, use this identifier with the operation.

startSentimentDetectionJobResponse_httpStatus :: Lens' StartSentimentDetectionJobResponse Int

The response's http status code.