Network.AWS.Comprehend.StopDominantLanguageDetectionJob

Description

Stops a dominant language detection job in progress.

If the job state is IN_PROGRESS the job is marked for termination and put into the STOP_REQUESTED state. If the job completes before it can be stopped, it is put into the COMPLETED state; otherwise the job is stopped and put into the STOPPED state.

If the job is in the COMPLETED or FAILED state when you call the StopDominantLanguageDetectionJob operation, the operation returns a 400 Internal Request Exception.

When a job is stopped, any documents already processed are written to the output location.

Creating a Request

data StopDominantLanguageDetectionJob

See: newStopDominantLanguageDetectionJob smart constructor.

Constructors

StopDominantLanguageDetectionJob'

Fields

jobId :: Text

The identifier of the dominant language detection job to stop.

newStopDominantLanguageDetectionJob

Arguments

:: Text
$sel:jobId:StopDominantLanguageDetectionJob'

-> StopDominantLanguageDetectionJob

Create a value of StopDominantLanguageDetectionJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobId:StopDominantLanguageDetectionJob', stopDominantLanguageDetectionJob_jobId - The identifier of the dominant language detection job to stop.

Request Lenses

stopDominantLanguageDetectionJob_jobId :: Lens' StopDominantLanguageDetectionJob Text

The identifier of the dominant language detection job to stop.

Destructuring the Response

data StopDominantLanguageDetectionJobResponse

See: newStopDominantLanguageDetectionJobResponse smart constructor.

Constructors

StopDominantLanguageDetectionJobResponse'

Fields

jobStatus :: Maybe JobStatus

Either STOP_REQUESTED if the job is currently running, or STOPPED if the job was previously stopped with the StopDominantLanguageDetectionJob operation.

jobId :: Maybe Text

The identifier of the dominant language detection job to stop.

httpStatus :: Int

The response's http status code.

newStopDominantLanguageDetectionJobResponse

Arguments

:: Int
$sel:httpStatus:StopDominantLanguageDetectionJobResponse'

-> StopDominantLanguageDetectionJobResponse

Create a value of StopDominantLanguageDetectionJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobStatus:StopDominantLanguageDetectionJobResponse', stopDominantLanguageDetectionJobResponse_jobStatus - Either STOP_REQUESTED if the job is currently running, or STOPPED if the job was previously stopped with the StopDominantLanguageDetectionJob operation.

$sel:jobId:StopDominantLanguageDetectionJob', stopDominantLanguageDetectionJobResponse_jobId - The identifier of the dominant language detection job to stop.

$sel:httpStatus:StopDominantLanguageDetectionJobResponse', stopDominantLanguageDetectionJobResponse_httpStatus - The response's http status code.

Response Lenses

stopDominantLanguageDetectionJobResponse_jobStatus :: Lens' StopDominantLanguageDetectionJobResponse (Maybe JobStatus)

Either STOP_REQUESTED if the job is currently running, or STOPPED if the job was previously stopped with the StopDominantLanguageDetectionJob operation.

stopDominantLanguageDetectionJobResponse_jobId :: Lens' StopDominantLanguageDetectionJobResponse (Maybe Text)

The identifier of the dominant language detection job to stop.

stopDominantLanguageDetectionJobResponse_httpStatus :: Lens' StopDominantLanguageDetectionJobResponse Int

The response's http status code.