Network.AWS.Comprehend.StopKeyPhrasesDetectionJob

Description

Stops a key phrases detection job in progress.

If the job state is IN_PROGRESS the job is marked for termination and put into the STOP_REQUESTED state. If the job completes before it can be stopped, it is put into the COMPLETED state; otherwise the job is stopped and put into the STOPPED state.

If the job is in the COMPLETED or FAILED state when you call the StopDominantLanguageDetectionJob operation, the operation returns a 400 Internal Request Exception.

When a job is stopped, any documents already processed are written to the output location.

Creating a Request

data StopKeyPhrasesDetectionJob

See: newStopKeyPhrasesDetectionJob smart constructor.

Constructors

StopKeyPhrasesDetectionJob'

Fields

jobId :: Text

The identifier of the key phrases detection job to stop.

newStopKeyPhrasesDetectionJob

Arguments

:: Text
$sel:jobId:StopKeyPhrasesDetectionJob'

-> StopKeyPhrasesDetectionJob

Create a value of StopKeyPhrasesDetectionJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobId:StopKeyPhrasesDetectionJob', stopKeyPhrasesDetectionJob_jobId - The identifier of the key phrases detection job to stop.

Request Lenses

stopKeyPhrasesDetectionJob_jobId :: Lens' StopKeyPhrasesDetectionJob Text

The identifier of the key phrases detection job to stop.

Destructuring the Response

data StopKeyPhrasesDetectionJobResponse

See: newStopKeyPhrasesDetectionJobResponse smart constructor.

Constructors

StopKeyPhrasesDetectionJobResponse'

Fields

jobStatus :: Maybe JobStatus

Either STOP_REQUESTED if the job is currently running, or STOPPED if the job was previously stopped with the StopKeyPhrasesDetectionJob operation.

jobId :: Maybe Text

The identifier of the key phrases detection job to stop.

httpStatus :: Int

The response's http status code.

newStopKeyPhrasesDetectionJobResponse

Arguments

:: Int
$sel:httpStatus:StopKeyPhrasesDetectionJobResponse'

-> StopKeyPhrasesDetectionJobResponse

Create a value of StopKeyPhrasesDetectionJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobStatus:StopKeyPhrasesDetectionJobResponse', stopKeyPhrasesDetectionJobResponse_jobStatus - Either STOP_REQUESTED if the job is currently running, or STOPPED if the job was previously stopped with the StopKeyPhrasesDetectionJob operation.

$sel:jobId:StopKeyPhrasesDetectionJob', stopKeyPhrasesDetectionJobResponse_jobId - The identifier of the key phrases detection job to stop.

$sel:httpStatus:StopKeyPhrasesDetectionJobResponse', stopKeyPhrasesDetectionJobResponse_httpStatus - The response's http status code.

Response Lenses

stopKeyPhrasesDetectionJobResponse_jobStatus :: Lens' StopKeyPhrasesDetectionJobResponse (Maybe JobStatus)

Either STOP_REQUESTED if the job is currently running, or STOPPED if the job was previously stopped with the StopKeyPhrasesDetectionJob operation.

stopKeyPhrasesDetectionJobResponse_jobId :: Lens' StopKeyPhrasesDetectionJobResponse (Maybe Text)

The identifier of the key phrases detection job to stop.

stopKeyPhrasesDetectionJobResponse_httpStatus :: Lens' StopKeyPhrasesDetectionJobResponse Int

The response's http status code.