Network.AWS.Comprehend.Types.AugmentedManifestsListItem

Description

Documentation

data AugmentedManifestsListItem

An augmented manifest file that provides training data for your custom model. An augmented manifest file is a labeled dataset that is produced by Amazon SageMaker Ground Truth.

See: newAugmentedManifestsListItem smart constructor.

Constructors

AugmentedManifestsListItem'

Fields

sourceDocumentsS3Uri :: Maybe Text

The S3 prefix to the source files (PDFs) that are referred to in the augmented manifest file.

documentType :: Maybe AugmentedManifestsDocumentTypeFormat

The type of augmented manifest. PlainTextDocument or SemiStructuredDocument. If you don't specify, the default is PlainTextDocument.

PLAIN_TEXT_DOCUMENT A document type that represents any unicode text that is encoded in UTF-8.

SEMI_STRUCTURED_DOCUMENT A document type with positional and structural context, like a PDF. For training with Amazon Comprehend, only PDFs are supported. For inference, Amazon Comprehend support PDFs, DOCX and TXT.

split :: Maybe Split

The purpose of the data you've provided in the augmented manifest. You can either train or test this data. If you don't specify, the default is train.

TRAIN - all of the documents in the manifest will be used for training. If no test documents are provided, Amazon Comprehend will automatically reserve a portion of the training documents for testing.

TEST - all of the documents in the manifest will be used for testing.

annotationDataS3Uri :: Maybe Text

The S3 prefix to the annotation files that are referred in the augmented manifest file.

s3Uri :: Text

The Amazon S3 location of the augmented manifest file.

attributeNames :: [Text]

The JSON attribute that contains the annotations for your training documents. The number of attribute names that you specify depends on whether your augmented manifest file is the output of a single labeling job or a chained labeling job.

If your file is the output of a single labeling job, specify the LabelAttributeName key that was used when the job was created in Ground Truth.

If your file is the output of a chained labeling job, specify the LabelAttributeName key for one or more jobs in the chain. Each LabelAttributeName key provides the annotations from an individual job.

newAugmentedManifestsListItem

Arguments

:: Text
$sel:s3Uri:AugmentedManifestsListItem'

-> AugmentedManifestsListItem

Create a value of AugmentedManifestsListItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceDocumentsS3Uri:AugmentedManifestsListItem', augmentedManifestsListItem_sourceDocumentsS3Uri - The S3 prefix to the source files (PDFs) that are referred to in the augmented manifest file.

$sel:documentType:AugmentedManifestsListItem', augmentedManifestsListItem_documentType - The type of augmented manifest. PlainTextDocument or SemiStructuredDocument. If you don't specify, the default is PlainTextDocument.

PLAIN_TEXT_DOCUMENT A document type that represents any unicode text that is encoded in UTF-8.

SEMI_STRUCTURED_DOCUMENT A document type with positional and structural context, like a PDF. For training with Amazon Comprehend, only PDFs are supported. For inference, Amazon Comprehend support PDFs, DOCX and TXT.

$sel:split:AugmentedManifestsListItem', augmentedManifestsListItem_split - The purpose of the data you've provided in the augmented manifest. You can either train or test this data. If you don't specify, the default is train.

TRAIN - all of the documents in the manifest will be used for training. If no test documents are provided, Amazon Comprehend will automatically reserve a portion of the training documents for testing.

TEST - all of the documents in the manifest will be used for testing.

$sel:annotationDataS3Uri:AugmentedManifestsListItem', augmentedManifestsListItem_annotationDataS3Uri - The S3 prefix to the annotation files that are referred in the augmented manifest file.

$sel:s3Uri:AugmentedManifestsListItem', augmentedManifestsListItem_s3Uri - The Amazon S3 location of the augmented manifest file.

$sel:attributeNames:AugmentedManifestsListItem', augmentedManifestsListItem_attributeNames - The JSON attribute that contains the annotations for your training documents. The number of attribute names that you specify depends on whether your augmented manifest file is the output of a single labeling job or a chained labeling job.

If your file is the output of a single labeling job, specify the LabelAttributeName key that was used when the job was created in Ground Truth.

If your file is the output of a chained labeling job, specify the LabelAttributeName key for one or more jobs in the chain. Each LabelAttributeName key provides the annotations from an individual job.

augmentedManifestsListItem_sourceDocumentsS3Uri :: Lens' AugmentedManifestsListItem (Maybe Text)

The S3 prefix to the source files (PDFs) that are referred to in the augmented manifest file.

augmentedManifestsListItem_documentType :: Lens' AugmentedManifestsListItem (Maybe AugmentedManifestsDocumentTypeFormat)

The type of augmented manifest. PlainTextDocument or SemiStructuredDocument. If you don't specify, the default is PlainTextDocument.

PLAIN_TEXT_DOCUMENT A document type that represents any unicode text that is encoded in UTF-8.

SEMI_STRUCTURED_DOCUMENT A document type with positional and structural context, like a PDF. For training with Amazon Comprehend, only PDFs are supported. For inference, Amazon Comprehend support PDFs, DOCX and TXT.

augmentedManifestsListItem_split :: Lens' AugmentedManifestsListItem (Maybe Split)

The purpose of the data you've provided in the augmented manifest. You can either train or test this data. If you don't specify, the default is train.

TRAIN - all of the documents in the manifest will be used for training. If no test documents are provided, Amazon Comprehend will automatically reserve a portion of the training documents for testing.

TEST - all of the documents in the manifest will be used for testing.

augmentedManifestsListItem_annotationDataS3Uri :: Lens' AugmentedManifestsListItem (Maybe Text)

The S3 prefix to the annotation files that are referred in the augmented manifest file.

augmentedManifestsListItem_s3Uri :: Lens' AugmentedManifestsListItem Text

The Amazon S3 location of the augmented manifest file.

augmentedManifestsListItem_attributeNames :: Lens' AugmentedManifestsListItem [Text]

The JSON attribute that contains the annotations for your training documents. The number of attribute names that you specify depends on whether your augmented manifest file is the output of a single labeling job or a chained labeling job.

If your file is the output of a single labeling job, specify the LabelAttributeName key that was used when the job was created in Ground Truth.

If your file is the output of a chained labeling job, specify the LabelAttributeName key for one or more jobs in the chain. Each LabelAttributeName key provides the annotations from an individual job.