Network.AWS.Comprehend.Types.BatchDetectSentimentItemResult

Description

Documentation

data BatchDetectSentimentItemResult

The result of calling the operation. The operation returns one object for each document that is successfully processed by the operation.

See: newBatchDetectSentimentItemResult smart constructor.

Constructors

BatchDetectSentimentItemResult'

Fields

sentiment :: Maybe SentimentType

The sentiment detected in the document.

sentimentScore :: Maybe SentimentScore

The level of confidence that Amazon Comprehend has in the accuracy of its sentiment detection.

index :: Maybe Int

The zero-based index of the document in the input list.

newBatchDetectSentimentItemResult :: BatchDetectSentimentItemResult

Create a value of BatchDetectSentimentItemResult with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sentiment:BatchDetectSentimentItemResult', batchDetectSentimentItemResult_sentiment - The sentiment detected in the document.

$sel:sentimentScore:BatchDetectSentimentItemResult', batchDetectSentimentItemResult_sentimentScore - The level of confidence that Amazon Comprehend has in the accuracy of its sentiment detection.

$sel:index:BatchDetectSentimentItemResult', batchDetectSentimentItemResult_index - The zero-based index of the document in the input list.

batchDetectSentimentItemResult_sentiment :: Lens' BatchDetectSentimentItemResult (Maybe SentimentType)

The sentiment detected in the document.

batchDetectSentimentItemResult_sentimentScore :: Lens' BatchDetectSentimentItemResult (Maybe SentimentScore)

The level of confidence that Amazon Comprehend has in the accuracy of its sentiment detection.

batchDetectSentimentItemResult_index :: Lens' BatchDetectSentimentItemResult (Maybe Int)

The zero-based index of the document in the input list.