Network.AWS.Comprehend.Types.ClassifierEvaluationMetrics

Description

Documentation

data ClassifierEvaluationMetrics

Describes the result metrics for the test data associated with an documentation classifier.

See: newClassifierEvaluationMetrics smart constructor.

Constructors

ClassifierEvaluationMetrics'

Fields

microPrecision :: Maybe Double

A measure of the usefulness of the recognizer results in the test data. High precision means that the recognizer returned substantially more relevant results than irrelevant ones. Unlike the Precision metric which comes from averaging the precision of all available labels, this is based on the overall score of all precision scores added together.

microF1Score :: Maybe Double

A measure of how accurate the classifier results are for the test data. It is a combination of the Micro Precision and Micro Recall values. The Micro F1Score is the harmonic mean of the two scores. The highest score is 1, and the worst score is 0.

recall :: Maybe Double

A measure of how complete the classifier results are for the test data. High recall means that the classifier returned most of the relevant results.

precision :: Maybe Double

A measure of the usefulness of the classifier results in the test data. High precision means that the classifier returned substantially more relevant results than irrelevant ones.

microRecall :: Maybe Double

A measure of how complete the classifier results are for the test data. High recall means that the classifier returned most of the relevant results. Specifically, this indicates how many of the correct categories in the text that the model can predict. It is a percentage of correct categories in the text that can found. Instead of averaging the recall scores of all labels (as with Recall), micro Recall is based on the overall score of all recall scores added together.

f1Score :: Maybe Double

A measure of how accurate the classifier results are for the test data. It is derived from the Precision and Recall values. The F1Score is the harmonic average of the two scores. The highest score is 1, and the worst score is 0.

hammingLoss :: Maybe Double

Indicates the fraction of labels that are incorrectly predicted. Also seen as the fraction of wrong labels compared to the total number of labels. Scores closer to zero are better.

accuracy :: Maybe Double

The fraction of the labels that were correct recognized. It is computed by dividing the number of labels in the test documents that were correctly recognized by the total number of labels in the test documents.

newClassifierEvaluationMetrics :: ClassifierEvaluationMetrics

Create a value of ClassifierEvaluationMetrics with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:microPrecision:ClassifierEvaluationMetrics', classifierEvaluationMetrics_microPrecision - A measure of the usefulness of the recognizer results in the test data. High precision means that the recognizer returned substantially more relevant results than irrelevant ones. Unlike the Precision metric which comes from averaging the precision of all available labels, this is based on the overall score of all precision scores added together.

$sel:microF1Score:ClassifierEvaluationMetrics', classifierEvaluationMetrics_microF1Score - A measure of how accurate the classifier results are for the test data. It is a combination of the Micro Precision and Micro Recall values. The Micro F1Score is the harmonic mean of the two scores. The highest score is 1, and the worst score is 0.

$sel:recall:ClassifierEvaluationMetrics', classifierEvaluationMetrics_recall - A measure of how complete the classifier results are for the test data. High recall means that the classifier returned most of the relevant results.

$sel:precision:ClassifierEvaluationMetrics', classifierEvaluationMetrics_precision - A measure of the usefulness of the classifier results in the test data. High precision means that the classifier returned substantially more relevant results than irrelevant ones.

$sel:microRecall:ClassifierEvaluationMetrics', classifierEvaluationMetrics_microRecall - A measure of how complete the classifier results are for the test data. High recall means that the classifier returned most of the relevant results. Specifically, this indicates how many of the correct categories in the text that the model can predict. It is a percentage of correct categories in the text that can found. Instead of averaging the recall scores of all labels (as with Recall), micro Recall is based on the overall score of all recall scores added together.

$sel:f1Score:ClassifierEvaluationMetrics', classifierEvaluationMetrics_f1Score - A measure of how accurate the classifier results are for the test data. It is derived from the Precision and Recall values. The F1Score is the harmonic average of the two scores. The highest score is 1, and the worst score is 0.

$sel:hammingLoss:ClassifierEvaluationMetrics', classifierEvaluationMetrics_hammingLoss - Indicates the fraction of labels that are incorrectly predicted. Also seen as the fraction of wrong labels compared to the total number of labels. Scores closer to zero are better.

$sel:accuracy:ClassifierEvaluationMetrics', classifierEvaluationMetrics_accuracy - The fraction of the labels that were correct recognized. It is computed by dividing the number of labels in the test documents that were correctly recognized by the total number of labels in the test documents.

classifierEvaluationMetrics_microPrecision :: Lens' ClassifierEvaluationMetrics (Maybe Double)

A measure of the usefulness of the recognizer results in the test data. High precision means that the recognizer returned substantially more relevant results than irrelevant ones. Unlike the Precision metric which comes from averaging the precision of all available labels, this is based on the overall score of all precision scores added together.

classifierEvaluationMetrics_microF1Score :: Lens' ClassifierEvaluationMetrics (Maybe Double)

A measure of how accurate the classifier results are for the test data. It is a combination of the Micro Precision and Micro Recall values. The Micro F1Score is the harmonic mean of the two scores. The highest score is 1, and the worst score is 0.

classifierEvaluationMetrics_recall :: Lens' ClassifierEvaluationMetrics (Maybe Double)

A measure of how complete the classifier results are for the test data. High recall means that the classifier returned most of the relevant results.

classifierEvaluationMetrics_precision :: Lens' ClassifierEvaluationMetrics (Maybe Double)

A measure of the usefulness of the classifier results in the test data. High precision means that the classifier returned substantially more relevant results than irrelevant ones.

classifierEvaluationMetrics_microRecall :: Lens' ClassifierEvaluationMetrics (Maybe Double)

A measure of how complete the classifier results are for the test data. High recall means that the classifier returned most of the relevant results. Specifically, this indicates how many of the correct categories in the text that the model can predict. It is a percentage of correct categories in the text that can found. Instead of averaging the recall scores of all labels (as with Recall), micro Recall is based on the overall score of all recall scores added together.

classifierEvaluationMetrics_f1Score :: Lens' ClassifierEvaluationMetrics (Maybe Double)

A measure of how accurate the classifier results are for the test data. It is derived from the Precision and Recall values. The F1Score is the harmonic average of the two scores. The highest score is 1, and the worst score is 0.

classifierEvaluationMetrics_hammingLoss :: Lens' ClassifierEvaluationMetrics (Maybe Double)

Indicates the fraction of labels that are incorrectly predicted. Also seen as the fraction of wrong labels compared to the total number of labels. Scores closer to zero are better.

classifierEvaluationMetrics_accuracy :: Lens' ClassifierEvaluationMetrics (Maybe Double)

The fraction of the labels that were correct recognized. It is computed by dividing the number of labels in the test documents that were correctly recognized by the total number of labels in the test documents.