Network.AWS.Comprehend.Types.DocumentClassificationJobProperties

Description

Documentation

data DocumentClassificationJobProperties

Provides information about a document classification job.

See: newDocumentClassificationJobProperties smart constructor.

Constructors

DocumentClassificationJobProperties'

Fields

vpcConfig :: Maybe VpcConfig

Configuration parameters for a private Virtual Private Cloud (VPC) containing the resources you are using for your document classification job. For more information, see Amazon VPC.

inputDataConfig :: Maybe InputDataConfig

The input data configuration that you supplied when you created the document classification job.

message :: Maybe Text

A description of the status of the job.

jobStatus :: Maybe JobStatus

The current status of the document classification job. If the status is FAILED, the Message field shows the reason for the failure.

outputDataConfig :: Maybe OutputDataConfig

The output data configuration that you supplied when you created the document classification job.

documentClassifierArn :: Maybe Text

The Amazon Resource Name (ARN) that identifies the document classifier.

endTime :: Maybe POSIX

The time that the document classification job completed.

jobArn :: Maybe Text

The Amazon Resource Name (ARN) of the document classification job. It is a unique, fully qualified identifier for the job. It includes the AWS account, Region, and the job ID. The format of the ARN is as follows:

arn:<partition>:comprehend:<region>:<account-id>:document-classification-job/<job-id>

The following is an example job ARN:

arn:aws:comprehend:us-west-2:111122223333:document-classification-job/1234abcd12ab34cd56ef1234567890ab

volumeKmsKeyId :: Maybe Text

ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt data on the storage volume attached to the ML compute instance(s) that process the analysis job. The VolumeKmsKeyId can be either of the following formats:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

submitTime :: Maybe POSIX

The time that the document classification job was submitted for processing.

jobName :: Maybe Text

The name that you assigned to the document classification job.

dataAccessRoleArn :: Maybe Text

The Amazon Resource Name (ARN) of the AWS identity and Access Management (IAM) role that grants Amazon Comprehend read access to your input data.

jobId :: Maybe Text

The identifier assigned to the document classification job.

newDocumentClassificationJobProperties :: DocumentClassificationJobProperties

Create a value of DocumentClassificationJobProperties with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpcConfig:DocumentClassificationJobProperties', documentClassificationJobProperties_vpcConfig - Configuration parameters for a private Virtual Private Cloud (VPC) containing the resources you are using for your document classification job. For more information, see Amazon VPC.

$sel:inputDataConfig:DocumentClassificationJobProperties', documentClassificationJobProperties_inputDataConfig - The input data configuration that you supplied when you created the document classification job.

$sel:message:DocumentClassificationJobProperties', documentClassificationJobProperties_message - A description of the status of the job.

$sel:jobStatus:DocumentClassificationJobProperties', documentClassificationJobProperties_jobStatus - The current status of the document classification job. If the status is FAILED, the Message field shows the reason for the failure.

$sel:outputDataConfig:DocumentClassificationJobProperties', documentClassificationJobProperties_outputDataConfig - The output data configuration that you supplied when you created the document classification job.

$sel:documentClassifierArn:DocumentClassificationJobProperties', documentClassificationJobProperties_documentClassifierArn - The Amazon Resource Name (ARN) that identifies the document classifier.

$sel:endTime:DocumentClassificationJobProperties', documentClassificationJobProperties_endTime - The time that the document classification job completed.

$sel:jobArn:DocumentClassificationJobProperties', documentClassificationJobProperties_jobArn - The Amazon Resource Name (ARN) of the document classification job. It is a unique, fully qualified identifier for the job. It includes the AWS account, Region, and the job ID. The format of the ARN is as follows:

arn:<partition>:comprehend:<region>:<account-id>:document-classification-job/<job-id>

The following is an example job ARN:

arn:aws:comprehend:us-west-2:111122223333:document-classification-job/1234abcd12ab34cd56ef1234567890ab

$sel:volumeKmsKeyId:DocumentClassificationJobProperties', documentClassificationJobProperties_volumeKmsKeyId - ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt data on the storage volume attached to the ML compute instance(s) that process the analysis job. The VolumeKmsKeyId can be either of the following formats:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

$sel:submitTime:DocumentClassificationJobProperties', documentClassificationJobProperties_submitTime - The time that the document classification job was submitted for processing.

$sel:jobName:DocumentClassificationJobProperties', documentClassificationJobProperties_jobName - The name that you assigned to the document classification job.

$sel:dataAccessRoleArn:DocumentClassificationJobProperties', documentClassificationJobProperties_dataAccessRoleArn - The Amazon Resource Name (ARN) of the AWS identity and Access Management (IAM) role that grants Amazon Comprehend read access to your input data.

$sel:jobId:DocumentClassificationJobProperties', documentClassificationJobProperties_jobId - The identifier assigned to the document classification job.

documentClassificationJobProperties_vpcConfig :: Lens' DocumentClassificationJobProperties (Maybe VpcConfig)

Configuration parameters for a private Virtual Private Cloud (VPC) containing the resources you are using for your document classification job. For more information, see Amazon VPC.

documentClassificationJobProperties_inputDataConfig :: Lens' DocumentClassificationJobProperties (Maybe InputDataConfig)

The input data configuration that you supplied when you created the document classification job.

documentClassificationJobProperties_message :: Lens' DocumentClassificationJobProperties (Maybe Text)

A description of the status of the job.

documentClassificationJobProperties_jobStatus :: Lens' DocumentClassificationJobProperties (Maybe JobStatus)

The current status of the document classification job. If the status is FAILED, the Message field shows the reason for the failure.

documentClassificationJobProperties_outputDataConfig :: Lens' DocumentClassificationJobProperties (Maybe OutputDataConfig)

The output data configuration that you supplied when you created the document classification job.

documentClassificationJobProperties_documentClassifierArn :: Lens' DocumentClassificationJobProperties (Maybe Text)

The Amazon Resource Name (ARN) that identifies the document classifier.

documentClassificationJobProperties_endTime :: Lens' DocumentClassificationJobProperties (Maybe UTCTime)

The time that the document classification job completed.

documentClassificationJobProperties_jobArn :: Lens' DocumentClassificationJobProperties (Maybe Text)

The Amazon Resource Name (ARN) of the document classification job. It is a unique, fully qualified identifier for the job. It includes the AWS account, Region, and the job ID. The format of the ARN is as follows:

arn:<partition>:comprehend:<region>:<account-id>:document-classification-job/<job-id>

The following is an example job ARN:

arn:aws:comprehend:us-west-2:111122223333:document-classification-job/1234abcd12ab34cd56ef1234567890ab

documentClassificationJobProperties_volumeKmsKeyId :: Lens' DocumentClassificationJobProperties (Maybe Text)

ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt data on the storage volume attached to the ML compute instance(s) that process the analysis job. The VolumeKmsKeyId can be either of the following formats:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

documentClassificationJobProperties_submitTime :: Lens' DocumentClassificationJobProperties (Maybe UTCTime)

The time that the document classification job was submitted for processing.

documentClassificationJobProperties_jobName :: Lens' DocumentClassificationJobProperties (Maybe Text)

The name that you assigned to the document classification job.

documentClassificationJobProperties_dataAccessRoleArn :: Lens' DocumentClassificationJobProperties (Maybe Text)

The Amazon Resource Name (ARN) of the AWS identity and Access Management (IAM) role that grants Amazon Comprehend read access to your input data.

documentClassificationJobProperties_jobId :: Lens' DocumentClassificationJobProperties (Maybe Text)

The identifier assigned to the document classification job.