Network.AWS.Comprehend.Types.DocumentClassifierProperties

Description

Documentation

data DocumentClassifierProperties

Provides information about a document classifier.

See: newDocumentClassifierProperties smart constructor.

Constructors

DocumentClassifierProperties'

Fields

status :: Maybe ModelStatus

The status of the document classifier. If the status is TRAINED the classifier is ready to use. If the status is FAILED you can see additional information about why the classifier wasn't trained in the Message field.

languageCode :: Maybe LanguageCode

The language code for the language of the documents that the classifier was trained on.

classifierMetadata :: Maybe (Sensitive ClassifierMetadata)

Information about the document classifier, including the number of documents used for training the classifier, the number of documents used for test the classifier, and an accuracy rating.

trainingEndTime :: Maybe POSIX

The time that training of the document classifier was completed. Indicates the time when the training completes on documentation classifiers. You are billed for the time interval between this time and the value of TrainingStartTime.

documentClassifierArn :: Maybe Text

The Amazon Resource Name (ARN) that identifies the document classifier.

versionName :: Maybe Text

The version name that you assigned to the document classifier.

modelKmsKeyId :: Maybe Text

ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt trained custom models. The ModelKmsKeyId can be either of the following formats:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

mode :: Maybe DocumentClassifierMode

Indicates the mode in which the specific classifier was trained. This also indicates the format of input documents and the format of the confusion matrix. Each classifier can only be trained in one mode and this cannot be changed once the classifier is trained.

inputDataConfig :: Maybe DocumentClassifierInputDataConfig

The input data configuration that you supplied when you created the document classifier for training.

vpcConfig :: Maybe VpcConfig

Configuration parameters for a private Virtual Private Cloud (VPC) containing the resources you are using for your custom classifier. For more information, see Amazon VPC.

volumeKmsKeyId :: Maybe Text

ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt data on the storage volume attached to the ML compute instance(s) that process the analysis job. The VolumeKmsKeyId can be either of the following formats:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

endTime :: Maybe POSIX

The time that training the document classifier completed.

outputDataConfig :: Maybe DocumentClassifierOutputDataConfig

Provides output results configuration parameters for custom classifier jobs.

trainingStartTime :: Maybe POSIX

Indicates the time when the training starts on documentation classifiers. You are billed for the time interval between this time and the value of TrainingEndTime.

dataAccessRoleArn :: Maybe Text

The Amazon Resource Name (ARN) of the AWS Identity and Management (IAM) role that grants Amazon Comprehend read access to your input data.

message :: Maybe Text

Additional information about the status of the classifier.

submitTime :: Maybe POSIX

The time that the document classifier was submitted for training.

newDocumentClassifierProperties :: DocumentClassifierProperties

Create a value of DocumentClassifierProperties with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DocumentClassifierProperties', documentClassifierProperties_status - The status of the document classifier. If the status is TRAINED the classifier is ready to use. If the status is FAILED you can see additional information about why the classifier wasn't trained in the Message field.

$sel:languageCode:DocumentClassifierProperties', documentClassifierProperties_languageCode - The language code for the language of the documents that the classifier was trained on.

$sel:classifierMetadata:DocumentClassifierProperties', documentClassifierProperties_classifierMetadata - Information about the document classifier, including the number of documents used for training the classifier, the number of documents used for test the classifier, and an accuracy rating.

$sel:trainingEndTime:DocumentClassifierProperties', documentClassifierProperties_trainingEndTime - The time that training of the document classifier was completed. Indicates the time when the training completes on documentation classifiers. You are billed for the time interval between this time and the value of TrainingStartTime.

$sel:documentClassifierArn:DocumentClassifierProperties', documentClassifierProperties_documentClassifierArn - The Amazon Resource Name (ARN) that identifies the document classifier.

$sel:versionName:DocumentClassifierProperties', documentClassifierProperties_versionName - The version name that you assigned to the document classifier.

$sel:modelKmsKeyId:DocumentClassifierProperties', documentClassifierProperties_modelKmsKeyId - ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt trained custom models. The ModelKmsKeyId can be either of the following formats:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

$sel:mode:DocumentClassifierProperties', documentClassifierProperties_mode - Indicates the mode in which the specific classifier was trained. This also indicates the format of input documents and the format of the confusion matrix. Each classifier can only be trained in one mode and this cannot be changed once the classifier is trained.

$sel:inputDataConfig:DocumentClassifierProperties', documentClassifierProperties_inputDataConfig - The input data configuration that you supplied when you created the document classifier for training.

$sel:vpcConfig:DocumentClassifierProperties', documentClassifierProperties_vpcConfig - Configuration parameters for a private Virtual Private Cloud (VPC) containing the resources you are using for your custom classifier. For more information, see Amazon VPC.

$sel:volumeKmsKeyId:DocumentClassifierProperties', documentClassifierProperties_volumeKmsKeyId - ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt data on the storage volume attached to the ML compute instance(s) that process the analysis job. The VolumeKmsKeyId can be either of the following formats:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

$sel:endTime:DocumentClassifierProperties', documentClassifierProperties_endTime - The time that training the document classifier completed.

$sel:outputDataConfig:DocumentClassifierProperties', documentClassifierProperties_outputDataConfig - Provides output results configuration parameters for custom classifier jobs.

$sel:trainingStartTime:DocumentClassifierProperties', documentClassifierProperties_trainingStartTime - Indicates the time when the training starts on documentation classifiers. You are billed for the time interval between this time and the value of TrainingEndTime.

$sel:dataAccessRoleArn:DocumentClassifierProperties', documentClassifierProperties_dataAccessRoleArn - The Amazon Resource Name (ARN) of the AWS Identity and Management (IAM) role that grants Amazon Comprehend read access to your input data.

$sel:message:DocumentClassifierProperties', documentClassifierProperties_message - Additional information about the status of the classifier.

$sel:submitTime:DocumentClassifierProperties', documentClassifierProperties_submitTime - The time that the document classifier was submitted for training.

documentClassifierProperties_status :: Lens' DocumentClassifierProperties (Maybe ModelStatus)

The status of the document classifier. If the status is TRAINED the classifier is ready to use. If the status is FAILED you can see additional information about why the classifier wasn't trained in the Message field.

documentClassifierProperties_languageCode :: Lens' DocumentClassifierProperties (Maybe LanguageCode)

The language code for the language of the documents that the classifier was trained on.

documentClassifierProperties_classifierMetadata :: Lens' DocumentClassifierProperties (Maybe ClassifierMetadata)

Information about the document classifier, including the number of documents used for training the classifier, the number of documents used for test the classifier, and an accuracy rating.

documentClassifierProperties_trainingEndTime :: Lens' DocumentClassifierProperties (Maybe UTCTime)

The time that training of the document classifier was completed. Indicates the time when the training completes on documentation classifiers. You are billed for the time interval between this time and the value of TrainingStartTime.

documentClassifierProperties_documentClassifierArn :: Lens' DocumentClassifierProperties (Maybe Text)

The Amazon Resource Name (ARN) that identifies the document classifier.

documentClassifierProperties_versionName :: Lens' DocumentClassifierProperties (Maybe Text)

The version name that you assigned to the document classifier.

documentClassifierProperties_modelKmsKeyId :: Lens' DocumentClassifierProperties (Maybe Text)

ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt trained custom models. The ModelKmsKeyId can be either of the following formats:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

documentClassifierProperties_mode :: Lens' DocumentClassifierProperties (Maybe DocumentClassifierMode)

Indicates the mode in which the specific classifier was trained. This also indicates the format of input documents and the format of the confusion matrix. Each classifier can only be trained in one mode and this cannot be changed once the classifier is trained.

documentClassifierProperties_inputDataConfig :: Lens' DocumentClassifierProperties (Maybe DocumentClassifierInputDataConfig)

The input data configuration that you supplied when you created the document classifier for training.

documentClassifierProperties_vpcConfig :: Lens' DocumentClassifierProperties (Maybe VpcConfig)

Configuration parameters for a private Virtual Private Cloud (VPC) containing the resources you are using for your custom classifier. For more information, see Amazon VPC.

documentClassifierProperties_volumeKmsKeyId :: Lens' DocumentClassifierProperties (Maybe Text)

ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt data on the storage volume attached to the ML compute instance(s) that process the analysis job. The VolumeKmsKeyId can be either of the following formats:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

documentClassifierProperties_endTime :: Lens' DocumentClassifierProperties (Maybe UTCTime)

The time that training the document classifier completed.

documentClassifierProperties_outputDataConfig :: Lens' DocumentClassifierProperties (Maybe DocumentClassifierOutputDataConfig)

Provides output results configuration parameters for custom classifier jobs.

documentClassifierProperties_trainingStartTime :: Lens' DocumentClassifierProperties (Maybe UTCTime)

Indicates the time when the training starts on documentation classifiers. You are billed for the time interval between this time and the value of TrainingEndTime.

documentClassifierProperties_dataAccessRoleArn :: Lens' DocumentClassifierProperties (Maybe Text)

The Amazon Resource Name (ARN) of the AWS Identity and Management (IAM) role that grants Amazon Comprehend read access to your input data.

documentClassifierProperties_message :: Lens' DocumentClassifierProperties (Maybe Text)

Additional information about the status of the classifier.

documentClassifierProperties_submitTime :: Lens' DocumentClassifierProperties (Maybe UTCTime)

The time that the document classifier was submitted for training.