Network.AWS.Comprehend.Types.DominantLanguageDetectionJobProperties

Description

Documentation

data DominantLanguageDetectionJobProperties

Provides information about a dominant language detection job.

See: newDominantLanguageDetectionJobProperties smart constructor.

Constructors

DominantLanguageDetectionJobProperties'

Fields

jobId :: Maybe Text

The identifier assigned to the dominant language detection job.

jobArn :: Maybe Text

The Amazon Resource Name (ARN) of the dominant language detection job. It is a unique, fully qualified identifier for the job. It includes the AWS account, Region, and the job ID. The format of the ARN is as follows:

arn:<partition>:comprehend:<region>:<account-id>:dominant-language-detection-job/<job-id>

The following is an example job ARN:

arn:aws:comprehend:us-west-2:111122223333:dominant-language-detection-job/1234abcd12ab34cd56ef1234567890ab

jobName :: Maybe Text

The name that you assigned to the dominant language detection job.

inputDataConfig :: Maybe InputDataConfig

The input data configuration that you supplied when you created the dominant language detection job.

vpcConfig :: Maybe VpcConfig

Configuration parameters for a private Virtual Private Cloud (VPC) containing the resources you are using for your dominant language detection job. For more information, see Amazon VPC.

volumeKmsKeyId :: Maybe Text

ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt data on the storage volume attached to the ML compute instance(s) that process the analysis job. The VolumeKmsKeyId can be either of the following formats:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

endTime :: Maybe POSIX

The time that the dominant language detection job completed.

outputDataConfig :: Maybe OutputDataConfig

The output data configuration that you supplied when you created the dominant language detection job.

dataAccessRoleArn :: Maybe Text

The Amazon Resource Name (ARN) that gives Amazon Comprehend read access to your input data.

jobStatus :: Maybe JobStatus

The current status of the dominant language detection job. If the status is FAILED, the Message field shows the reason for the failure.

message :: Maybe Text

A description for the status of a job.

submitTime :: Maybe POSIX

The time that the dominant language detection job was submitted for processing.

newDominantLanguageDetectionJobProperties :: DominantLanguageDetectionJobProperties

Create a value of DominantLanguageDetectionJobProperties with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobId:DominantLanguageDetectionJobProperties', dominantLanguageDetectionJobProperties_jobId - The identifier assigned to the dominant language detection job.

$sel:jobArn:DominantLanguageDetectionJobProperties', dominantLanguageDetectionJobProperties_jobArn - The Amazon Resource Name (ARN) of the dominant language detection job. It is a unique, fully qualified identifier for the job. It includes the AWS account, Region, and the job ID. The format of the ARN is as follows:

arn:<partition>:comprehend:<region>:<account-id>:dominant-language-detection-job/<job-id>

The following is an example job ARN:

arn:aws:comprehend:us-west-2:111122223333:dominant-language-detection-job/1234abcd12ab34cd56ef1234567890ab

$sel:jobName:DominantLanguageDetectionJobProperties', dominantLanguageDetectionJobProperties_jobName - The name that you assigned to the dominant language detection job.

$sel:inputDataConfig:DominantLanguageDetectionJobProperties', dominantLanguageDetectionJobProperties_inputDataConfig - The input data configuration that you supplied when you created the dominant language detection job.

$sel:vpcConfig:DominantLanguageDetectionJobProperties', dominantLanguageDetectionJobProperties_vpcConfig - Configuration parameters for a private Virtual Private Cloud (VPC) containing the resources you are using for your dominant language detection job. For more information, see Amazon VPC.

$sel:volumeKmsKeyId:DominantLanguageDetectionJobProperties', dominantLanguageDetectionJobProperties_volumeKmsKeyId - ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt data on the storage volume attached to the ML compute instance(s) that process the analysis job. The VolumeKmsKeyId can be either of the following formats:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

$sel:endTime:DominantLanguageDetectionJobProperties', dominantLanguageDetectionJobProperties_endTime - The time that the dominant language detection job completed.

$sel:outputDataConfig:DominantLanguageDetectionJobProperties', dominantLanguageDetectionJobProperties_outputDataConfig - The output data configuration that you supplied when you created the dominant language detection job.

$sel:dataAccessRoleArn:DominantLanguageDetectionJobProperties', dominantLanguageDetectionJobProperties_dataAccessRoleArn - The Amazon Resource Name (ARN) that gives Amazon Comprehend read access to your input data.

$sel:jobStatus:DominantLanguageDetectionJobProperties', dominantLanguageDetectionJobProperties_jobStatus - The current status of the dominant language detection job. If the status is FAILED, the Message field shows the reason for the failure.

$sel:message:DominantLanguageDetectionJobProperties', dominantLanguageDetectionJobProperties_message - A description for the status of a job.

$sel:submitTime:DominantLanguageDetectionJobProperties', dominantLanguageDetectionJobProperties_submitTime - The time that the dominant language detection job was submitted for processing.

dominantLanguageDetectionJobProperties_jobId :: Lens' DominantLanguageDetectionJobProperties (Maybe Text)

The identifier assigned to the dominant language detection job.

dominantLanguageDetectionJobProperties_jobArn :: Lens' DominantLanguageDetectionJobProperties (Maybe Text)

The Amazon Resource Name (ARN) of the dominant language detection job. It is a unique, fully qualified identifier for the job. It includes the AWS account, Region, and the job ID. The format of the ARN is as follows:

arn:<partition>:comprehend:<region>:<account-id>:dominant-language-detection-job/<job-id>

The following is an example job ARN:

arn:aws:comprehend:us-west-2:111122223333:dominant-language-detection-job/1234abcd12ab34cd56ef1234567890ab

dominantLanguageDetectionJobProperties_jobName :: Lens' DominantLanguageDetectionJobProperties (Maybe Text)

The name that you assigned to the dominant language detection job.

dominantLanguageDetectionJobProperties_inputDataConfig :: Lens' DominantLanguageDetectionJobProperties (Maybe InputDataConfig)

The input data configuration that you supplied when you created the dominant language detection job.

dominantLanguageDetectionJobProperties_vpcConfig :: Lens' DominantLanguageDetectionJobProperties (Maybe VpcConfig)

Configuration parameters for a private Virtual Private Cloud (VPC) containing the resources you are using for your dominant language detection job. For more information, see Amazon VPC.

dominantLanguageDetectionJobProperties_volumeKmsKeyId :: Lens' DominantLanguageDetectionJobProperties (Maybe Text)

ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt data on the storage volume attached to the ML compute instance(s) that process the analysis job. The VolumeKmsKeyId can be either of the following formats:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

dominantLanguageDetectionJobProperties_endTime :: Lens' DominantLanguageDetectionJobProperties (Maybe UTCTime)

The time that the dominant language detection job completed.

dominantLanguageDetectionJobProperties_outputDataConfig :: Lens' DominantLanguageDetectionJobProperties (Maybe OutputDataConfig)

The output data configuration that you supplied when you created the dominant language detection job.

dominantLanguageDetectionJobProperties_dataAccessRoleArn :: Lens' DominantLanguageDetectionJobProperties (Maybe Text)

The Amazon Resource Name (ARN) that gives Amazon Comprehend read access to your input data.

dominantLanguageDetectionJobProperties_jobStatus :: Lens' DominantLanguageDetectionJobProperties (Maybe JobStatus)

The current status of the dominant language detection job. If the status is FAILED, the Message field shows the reason for the failure.

dominantLanguageDetectionJobProperties_message :: Lens' DominantLanguageDetectionJobProperties (Maybe Text)

A description for the status of a job.

dominantLanguageDetectionJobProperties_submitTime :: Lens' DominantLanguageDetectionJobProperties (Maybe UTCTime)

The time that the dominant language detection job was submitted for processing.