Network.AWS.Comprehend.Types.EndpointProperties

Description

Documentation

data EndpointProperties

Specifies information about the specified endpoint.

See: newEndpointProperties smart constructor.

Constructors

EndpointProperties'

Fields

creationTime :: Maybe POSIX

The creation date and time of the endpoint.

status :: Maybe EndpointStatus

Specifies the status of the endpoint. Because the endpoint updates and creation are asynchronous, so customers will need to wait for the endpoint to be Ready status before making inference requests.

modelArn :: Maybe Text

The Amazon Resource Number (ARN) of the model to which the endpoint is attached.

desiredModelArn :: Maybe Text

ARN of the new model to use for updating an existing endpoint. This ARN is going to be different from the model ARN when the update is in progress

lastModifiedTime :: Maybe POSIX

The date and time that the endpoint was last modified.

desiredInferenceUnits :: Maybe Natural

The desired number of inference units to be used by the model using this endpoint. Each inference unit represents of a throughput of 100 characters per second.

currentInferenceUnits :: Maybe Natural

The number of inference units currently used by the model using this endpoint.

dataAccessRoleArn :: Maybe Text

The Amazon Resource Name (ARN) of the AWS identity and Access Management (IAM) role that grants Amazon Comprehend read access to trained custom models encrypted with a customer managed key (ModelKmsKeyId).

desiredDataAccessRoleArn :: Maybe Text

Data access role ARN to use in case the new model is encrypted with a customer KMS key.

message :: Maybe Text

Specifies a reason for failure in cases of Failed status.

endpointArn :: Maybe Text

The Amazon Resource Number (ARN) of the endpoint.

newEndpointProperties :: EndpointProperties

Create a value of EndpointProperties with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:EndpointProperties', endpointProperties_creationTime - The creation date and time of the endpoint.

$sel:status:EndpointProperties', endpointProperties_status - Specifies the status of the endpoint. Because the endpoint updates and creation are asynchronous, so customers will need to wait for the endpoint to be Ready status before making inference requests.

$sel:modelArn:EndpointProperties', endpointProperties_modelArn - The Amazon Resource Number (ARN) of the model to which the endpoint is attached.

$sel:desiredModelArn:EndpointProperties', endpointProperties_desiredModelArn - ARN of the new model to use for updating an existing endpoint. This ARN is going to be different from the model ARN when the update is in progress

$sel:lastModifiedTime:EndpointProperties', endpointProperties_lastModifiedTime - The date and time that the endpoint was last modified.

$sel:desiredInferenceUnits:EndpointProperties', endpointProperties_desiredInferenceUnits - The desired number of inference units to be used by the model using this endpoint. Each inference unit represents of a throughput of 100 characters per second.

$sel:currentInferenceUnits:EndpointProperties', endpointProperties_currentInferenceUnits - The number of inference units currently used by the model using this endpoint.

$sel:dataAccessRoleArn:EndpointProperties', endpointProperties_dataAccessRoleArn - The Amazon Resource Name (ARN) of the AWS identity and Access Management (IAM) role that grants Amazon Comprehend read access to trained custom models encrypted with a customer managed key (ModelKmsKeyId).

$sel:desiredDataAccessRoleArn:EndpointProperties', endpointProperties_desiredDataAccessRoleArn - Data access role ARN to use in case the new model is encrypted with a customer KMS key.

$sel:message:EndpointProperties', endpointProperties_message - Specifies a reason for failure in cases of Failed status.

$sel:endpointArn:EndpointProperties', endpointProperties_endpointArn - The Amazon Resource Number (ARN) of the endpoint.

endpointProperties_creationTime :: Lens' EndpointProperties (Maybe UTCTime)

The creation date and time of the endpoint.

endpointProperties_status :: Lens' EndpointProperties (Maybe EndpointStatus)

Specifies the status of the endpoint. Because the endpoint updates and creation are asynchronous, so customers will need to wait for the endpoint to be Ready status before making inference requests.

endpointProperties_modelArn :: Lens' EndpointProperties (Maybe Text)

The Amazon Resource Number (ARN) of the model to which the endpoint is attached.

endpointProperties_desiredModelArn :: Lens' EndpointProperties (Maybe Text)

ARN of the new model to use for updating an existing endpoint. This ARN is going to be different from the model ARN when the update is in progress

endpointProperties_lastModifiedTime :: Lens' EndpointProperties (Maybe UTCTime)

The date and time that the endpoint was last modified.

endpointProperties_desiredInferenceUnits :: Lens' EndpointProperties (Maybe Natural)

The desired number of inference units to be used by the model using this endpoint. Each inference unit represents of a throughput of 100 characters per second.

endpointProperties_currentInferenceUnits :: Lens' EndpointProperties (Maybe Natural)

The number of inference units currently used by the model using this endpoint.

endpointProperties_dataAccessRoleArn :: Lens' EndpointProperties (Maybe Text)

The Amazon Resource Name (ARN) of the AWS identity and Access Management (IAM) role that grants Amazon Comprehend read access to trained custom models encrypted with a customer managed key (ModelKmsKeyId).

endpointProperties_desiredDataAccessRoleArn :: Lens' EndpointProperties (Maybe Text)

Data access role ARN to use in case the new model is encrypted with a customer KMS key.

endpointProperties_message :: Lens' EndpointProperties (Maybe Text)

Specifies a reason for failure in cases of Failed status.

endpointProperties_endpointArn :: Lens' EndpointProperties (Maybe Text)

The Amazon Resource Number (ARN) of the endpoint.