Network.AWS.Comprehend.Types.EntitiesDetectionJobProperties

Description

Documentation

data EntitiesDetectionJobProperties

Provides information about an entities detection job.

See: newEntitiesDetectionJobProperties smart constructor.

Constructors

EntitiesDetectionJobProperties'

Fields

languageCode :: Maybe LanguageCode

The language code of the input documents.

vpcConfig :: Maybe VpcConfig

Configuration parameters for a private Virtual Private Cloud (VPC) containing the resources you are using for your entity detection job. For more information, see Amazon VPC.

inputDataConfig :: Maybe InputDataConfig

The input data configuration that you supplied when you created the entities detection job.

message :: Maybe Text

A description of the status of a job.

jobStatus :: Maybe JobStatus

The current status of the entities detection job. If the status is FAILED, the Message field shows the reason for the failure.

outputDataConfig :: Maybe OutputDataConfig

The output data configuration that you supplied when you created the entities detection job.

endTime :: Maybe POSIX

The time that the entities detection job completed

jobArn :: Maybe Text

The Amazon Resource Name (ARN) of the entities detection job. It is a unique, fully qualified identifier for the job. It includes the AWS account, Region, and the job ID. The format of the ARN is as follows:

arn:<partition>:comprehend:<region>:<account-id>:entities-detection-job/<job-id>

The following is an example job ARN:

arn:aws:comprehend:us-west-2:111122223333:entities-detection-job/1234abcd12ab34cd56ef1234567890ab

volumeKmsKeyId :: Maybe Text

ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt data on the storage volume attached to the ML compute instance(s) that process the analysis job. The VolumeKmsKeyId can be either of the following formats:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

submitTime :: Maybe POSIX

The time that the entities detection job was submitted for processing.

entityRecognizerArn :: Maybe Text

The Amazon Resource Name (ARN) that identifies the entity recognizer.

jobName :: Maybe Text

The name that you assigned the entities detection job.

dataAccessRoleArn :: Maybe Text

The Amazon Resource Name (ARN) that gives Amazon Comprehend read access to your input data.

jobId :: Maybe Text

The identifier assigned to the entities detection job.

newEntitiesDetectionJobProperties :: EntitiesDetectionJobProperties

Create a value of EntitiesDetectionJobProperties with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:languageCode:EntitiesDetectionJobProperties', entitiesDetectionJobProperties_languageCode - The language code of the input documents.

$sel:vpcConfig:EntitiesDetectionJobProperties', entitiesDetectionJobProperties_vpcConfig - Configuration parameters for a private Virtual Private Cloud (VPC) containing the resources you are using for your entity detection job. For more information, see Amazon VPC.

$sel:inputDataConfig:EntitiesDetectionJobProperties', entitiesDetectionJobProperties_inputDataConfig - The input data configuration that you supplied when you created the entities detection job.

$sel:message:EntitiesDetectionJobProperties', entitiesDetectionJobProperties_message - A description of the status of a job.

$sel:jobStatus:EntitiesDetectionJobProperties', entitiesDetectionJobProperties_jobStatus - The current status of the entities detection job. If the status is FAILED, the Message field shows the reason for the failure.

$sel:outputDataConfig:EntitiesDetectionJobProperties', entitiesDetectionJobProperties_outputDataConfig - The output data configuration that you supplied when you created the entities detection job.

$sel:endTime:EntitiesDetectionJobProperties', entitiesDetectionJobProperties_endTime - The time that the entities detection job completed

$sel:jobArn:EntitiesDetectionJobProperties', entitiesDetectionJobProperties_jobArn - The Amazon Resource Name (ARN) of the entities detection job. It is a unique, fully qualified identifier for the job. It includes the AWS account, Region, and the job ID. The format of the ARN is as follows:

arn:<partition>:comprehend:<region>:<account-id>:entities-detection-job/<job-id>

The following is an example job ARN:

arn:aws:comprehend:us-west-2:111122223333:entities-detection-job/1234abcd12ab34cd56ef1234567890ab

$sel:volumeKmsKeyId:EntitiesDetectionJobProperties', entitiesDetectionJobProperties_volumeKmsKeyId - ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt data on the storage volume attached to the ML compute instance(s) that process the analysis job. The VolumeKmsKeyId can be either of the following formats:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

$sel:submitTime:EntitiesDetectionJobProperties', entitiesDetectionJobProperties_submitTime - The time that the entities detection job was submitted for processing.

$sel:entityRecognizerArn:EntitiesDetectionJobProperties', entitiesDetectionJobProperties_entityRecognizerArn - The Amazon Resource Name (ARN) that identifies the entity recognizer.

$sel:jobName:EntitiesDetectionJobProperties', entitiesDetectionJobProperties_jobName - The name that you assigned the entities detection job.

$sel:dataAccessRoleArn:EntitiesDetectionJobProperties', entitiesDetectionJobProperties_dataAccessRoleArn - The Amazon Resource Name (ARN) that gives Amazon Comprehend read access to your input data.

$sel:jobId:EntitiesDetectionJobProperties', entitiesDetectionJobProperties_jobId - The identifier assigned to the entities detection job.

entitiesDetectionJobProperties_languageCode :: Lens' EntitiesDetectionJobProperties (Maybe LanguageCode)

The language code of the input documents.

entitiesDetectionJobProperties_vpcConfig :: Lens' EntitiesDetectionJobProperties (Maybe VpcConfig)

Configuration parameters for a private Virtual Private Cloud (VPC) containing the resources you are using for your entity detection job. For more information, see Amazon VPC.

entitiesDetectionJobProperties_inputDataConfig :: Lens' EntitiesDetectionJobProperties (Maybe InputDataConfig)

The input data configuration that you supplied when you created the entities detection job.

entitiesDetectionJobProperties_message :: Lens' EntitiesDetectionJobProperties (Maybe Text)

A description of the status of a job.

entitiesDetectionJobProperties_jobStatus :: Lens' EntitiesDetectionJobProperties (Maybe JobStatus)

The current status of the entities detection job. If the status is FAILED, the Message field shows the reason for the failure.

entitiesDetectionJobProperties_outputDataConfig :: Lens' EntitiesDetectionJobProperties (Maybe OutputDataConfig)

The output data configuration that you supplied when you created the entities detection job.

entitiesDetectionJobProperties_endTime :: Lens' EntitiesDetectionJobProperties (Maybe UTCTime)

The time that the entities detection job completed

entitiesDetectionJobProperties_jobArn :: Lens' EntitiesDetectionJobProperties (Maybe Text)

The Amazon Resource Name (ARN) of the entities detection job. It is a unique, fully qualified identifier for the job. It includes the AWS account, Region, and the job ID. The format of the ARN is as follows:

arn:<partition>:comprehend:<region>:<account-id>:entities-detection-job/<job-id>

The following is an example job ARN:

arn:aws:comprehend:us-west-2:111122223333:entities-detection-job/1234abcd12ab34cd56ef1234567890ab

entitiesDetectionJobProperties_volumeKmsKeyId :: Lens' EntitiesDetectionJobProperties (Maybe Text)

ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt data on the storage volume attached to the ML compute instance(s) that process the analysis job. The VolumeKmsKeyId can be either of the following formats:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

entitiesDetectionJobProperties_submitTime :: Lens' EntitiesDetectionJobProperties (Maybe UTCTime)

The time that the entities detection job was submitted for processing.

entitiesDetectionJobProperties_entityRecognizerArn :: Lens' EntitiesDetectionJobProperties (Maybe Text)

The Amazon Resource Name (ARN) that identifies the entity recognizer.

entitiesDetectionJobProperties_jobName :: Lens' EntitiesDetectionJobProperties (Maybe Text)

The name that you assigned the entities detection job.

entitiesDetectionJobProperties_dataAccessRoleArn :: Lens' EntitiesDetectionJobProperties (Maybe Text)

The Amazon Resource Name (ARN) that gives Amazon Comprehend read access to your input data.

entitiesDetectionJobProperties_jobId :: Lens' EntitiesDetectionJobProperties (Maybe Text)

The identifier assigned to the entities detection job.