Network.AWS.Comprehend.Types.EntityRecognizerEvaluationMetrics

Description

Documentation

data EntityRecognizerEvaluationMetrics

Detailed information about the accuracy of an entity recognizer.

See: newEntityRecognizerEvaluationMetrics smart constructor.

Constructors

EntityRecognizerEvaluationMetrics'

Fields

f1Score :: Maybe Double

A measure of how accurate the recognizer results are for the test data. It is derived from the Precision and Recall values. The F1Score is the harmonic average of the two scores. The highest score is 1, and the worst score is 0.

precision :: Maybe Double

A measure of the usefulness of the recognizer results in the test data. High precision means that the recognizer returned substantially more relevant results than irrelevant ones.

recall :: Maybe Double

A measure of how complete the recognizer results are for the test data. High recall means that the recognizer returned most of the relevant results.

newEntityRecognizerEvaluationMetrics :: EntityRecognizerEvaluationMetrics

Create a value of EntityRecognizerEvaluationMetrics with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:f1Score:EntityRecognizerEvaluationMetrics', entityRecognizerEvaluationMetrics_f1Score - A measure of how accurate the recognizer results are for the test data. It is derived from the Precision and Recall values. The F1Score is the harmonic average of the two scores. The highest score is 1, and the worst score is 0.

$sel:precision:EntityRecognizerEvaluationMetrics', entityRecognizerEvaluationMetrics_precision - A measure of the usefulness of the recognizer results in the test data. High precision means that the recognizer returned substantially more relevant results than irrelevant ones.

$sel:recall:EntityRecognizerEvaluationMetrics', entityRecognizerEvaluationMetrics_recall - A measure of how complete the recognizer results are for the test data. High recall means that the recognizer returned most of the relevant results.

entityRecognizerEvaluationMetrics_f1Score :: Lens' EntityRecognizerEvaluationMetrics (Maybe Double)

A measure of how accurate the recognizer results are for the test data. It is derived from the Precision and Recall values. The F1Score is the harmonic average of the two scores. The highest score is 1, and the worst score is 0.

entityRecognizerEvaluationMetrics_precision :: Lens' EntityRecognizerEvaluationMetrics (Maybe Double)

A measure of the usefulness of the recognizer results in the test data. High precision means that the recognizer returned substantially more relevant results than irrelevant ones.

entityRecognizerEvaluationMetrics_recall :: Lens' EntityRecognizerEvaluationMetrics (Maybe Double)

A measure of how complete the recognizer results are for the test data. High recall means that the recognizer returned most of the relevant results.