Network.AWS.Comprehend.Types.EntityRecognizerInputDataConfig

Description

Documentation

data EntityRecognizerInputDataConfig

Specifies the format and location of the input data.

See: newEntityRecognizerInputDataConfig smart constructor.

Constructors

EntityRecognizerInputDataConfig'

Fields

augmentedManifests :: Maybe [AugmentedManifestsListItem]

A list of augmented manifest files that provide training data for your custom model. An augmented manifest file is a labeled dataset that is produced by Amazon SageMaker Ground Truth.

This parameter is required if you set DataFormat to AUGMENTED_MANIFEST.

annotations :: Maybe EntityRecognizerAnnotations

The S3 location of the CSV file that annotates your training documents.

dataFormat :: Maybe EntityRecognizerDataFormat

The format of your training data:

COMPREHEND_CSV: A CSV file that supplements your training documents. The CSV file contains information about the custom entities that your trained model will detect. The required format of the file depends on whether you are providing annotations or an entity list.

If you use this value, you must provide your CSV file by using either the Annotations or EntityList parameters. You must provide your training documents by using the Documents parameter.

AUGMENTED_MANIFEST: A labeled dataset that is produced by Amazon SageMaker Ground Truth. This file is in JSON lines format. Each line is a complete JSON object that contains a training document and its labels. Each label annotates a named entity in the training document.

If you use this value, you must provide the AugmentedManifests parameter in your request.

If you don't specify a value, Amazon Comprehend uses COMPREHEND_CSV as the default.

documents :: Maybe EntityRecognizerDocuments

The S3 location of the folder that contains the training documents for your custom entity recognizer.

This parameter is required if you set DataFormat to COMPREHEND_CSV.

entityList :: Maybe EntityRecognizerEntityList

The S3 location of the CSV file that has the entity list for your custom entity recognizer.

entityTypes :: [EntityTypesListItem]

The entity types in the labeled training data that Amazon Comprehend uses to train the custom entity recognizer. Any entity types that you don't specify are ignored.

A maximum of 25 entity types can be used at one time to train an entity recognizer. Entity types must not contain the following invalid characters: \n (line break), \\n (escaped line break), \r (carriage return), \\r (escaped carriage return), \t (tab), \\t (escaped tab), space, and , (comma).

newEntityRecognizerInputDataConfig :: EntityRecognizerInputDataConfig

Create a value of EntityRecognizerInputDataConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:augmentedManifests:EntityRecognizerInputDataConfig', entityRecognizerInputDataConfig_augmentedManifests - A list of augmented manifest files that provide training data for your custom model. An augmented manifest file is a labeled dataset that is produced by Amazon SageMaker Ground Truth.

This parameter is required if you set DataFormat to AUGMENTED_MANIFEST.

$sel:annotations:EntityRecognizerInputDataConfig', entityRecognizerInputDataConfig_annotations - The S3 location of the CSV file that annotates your training documents.

$sel:dataFormat:EntityRecognizerInputDataConfig', entityRecognizerInputDataConfig_dataFormat - The format of your training data:

COMPREHEND_CSV: A CSV file that supplements your training documents. The CSV file contains information about the custom entities that your trained model will detect. The required format of the file depends on whether you are providing annotations or an entity list.

If you use this value, you must provide your CSV file by using either the Annotations or EntityList parameters. You must provide your training documents by using the Documents parameter.

AUGMENTED_MANIFEST: A labeled dataset that is produced by Amazon SageMaker Ground Truth. This file is in JSON lines format. Each line is a complete JSON object that contains a training document and its labels. Each label annotates a named entity in the training document.

If you use this value, you must provide the AugmentedManifests parameter in your request.

If you don't specify a value, Amazon Comprehend uses COMPREHEND_CSV as the default.

$sel:documents:EntityRecognizerInputDataConfig', entityRecognizerInputDataConfig_documents - The S3 location of the folder that contains the training documents for your custom entity recognizer.

This parameter is required if you set DataFormat to COMPREHEND_CSV.

$sel:entityList:EntityRecognizerInputDataConfig', entityRecognizerInputDataConfig_entityList - The S3 location of the CSV file that has the entity list for your custom entity recognizer.

$sel:entityTypes:EntityRecognizerInputDataConfig', entityRecognizerInputDataConfig_entityTypes - The entity types in the labeled training data that Amazon Comprehend uses to train the custom entity recognizer. Any entity types that you don't specify are ignored.

A maximum of 25 entity types can be used at one time to train an entity recognizer. Entity types must not contain the following invalid characters: \n (line break), \\n (escaped line break), \r (carriage return), \\r (escaped carriage return), \t (tab), \\t (escaped tab), space, and , (comma).

entityRecognizerInputDataConfig_augmentedManifests :: Lens' EntityRecognizerInputDataConfig (Maybe [AugmentedManifestsListItem])

A list of augmented manifest files that provide training data for your custom model. An augmented manifest file is a labeled dataset that is produced by Amazon SageMaker Ground Truth.

This parameter is required if you set DataFormat to AUGMENTED_MANIFEST.

entityRecognizerInputDataConfig_annotations :: Lens' EntityRecognizerInputDataConfig (Maybe EntityRecognizerAnnotations)

The S3 location of the CSV file that annotates your training documents.

entityRecognizerInputDataConfig_dataFormat :: Lens' EntityRecognizerInputDataConfig (Maybe EntityRecognizerDataFormat)

The format of your training data:

COMPREHEND_CSV: A CSV file that supplements your training documents. The CSV file contains information about the custom entities that your trained model will detect. The required format of the file depends on whether you are providing annotations or an entity list.

If you use this value, you must provide your CSV file by using either the Annotations or EntityList parameters. You must provide your training documents by using the Documents parameter.

AUGMENTED_MANIFEST: A labeled dataset that is produced by Amazon SageMaker Ground Truth. This file is in JSON lines format. Each line is a complete JSON object that contains a training document and its labels. Each label annotates a named entity in the training document.

If you use this value, you must provide the AugmentedManifests parameter in your request.

If you don't specify a value, Amazon Comprehend uses COMPREHEND_CSV as the default.

entityRecognizerInputDataConfig_documents :: Lens' EntityRecognizerInputDataConfig (Maybe EntityRecognizerDocuments)

The S3 location of the folder that contains the training documents for your custom entity recognizer.

This parameter is required if you set DataFormat to COMPREHEND_CSV.

entityRecognizerInputDataConfig_entityList :: Lens' EntityRecognizerInputDataConfig (Maybe EntityRecognizerEntityList)

The S3 location of the CSV file that has the entity list for your custom entity recognizer.

entityRecognizerInputDataConfig_entityTypes :: Lens' EntityRecognizerInputDataConfig [EntityTypesListItem]

The entity types in the labeled training data that Amazon Comprehend uses to train the custom entity recognizer. Any entity types that you don't specify are ignored.

A maximum of 25 entity types can be used at one time to train an entity recognizer. Entity types must not contain the following invalid characters: \n (line break), \\n (escaped line break), \r (carriage return), \\r (escaped carriage return), \t (tab), \\t (escaped tab), space, and , (comma).