Network.AWS.Comprehend.Types.EntityRecognizerMetadata

Description

Documentation

data EntityRecognizerMetadata

Detailed information about an entity recognizer.

See: newEntityRecognizerMetadata smart constructor.

Constructors

EntityRecognizerMetadata'

Fields

numberOfTestDocuments :: Maybe Int

The number of documents in the input data that were used to test the entity recognizer. Typically this is 10 to 20 percent of the input documents.

numberOfTrainedDocuments :: Maybe Int

The number of documents in the input data that were used to train the entity recognizer. Typically this is 80 to 90 percent of the input documents.

evaluationMetrics :: Maybe EntityRecognizerEvaluationMetrics

Detailed information about the accuracy of an entity recognizer.

entityTypes :: Maybe [EntityRecognizerMetadataEntityTypesListItem]

Entity types from the metadata of an entity recognizer.

newEntityRecognizerMetadata :: EntityRecognizerMetadata

Create a value of EntityRecognizerMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:numberOfTestDocuments:EntityRecognizerMetadata', entityRecognizerMetadata_numberOfTestDocuments - The number of documents in the input data that were used to test the entity recognizer. Typically this is 10 to 20 percent of the input documents.

$sel:numberOfTrainedDocuments:EntityRecognizerMetadata', entityRecognizerMetadata_numberOfTrainedDocuments - The number of documents in the input data that were used to train the entity recognizer. Typically this is 80 to 90 percent of the input documents.

$sel:evaluationMetrics:EntityRecognizerMetadata', entityRecognizerMetadata_evaluationMetrics - Detailed information about the accuracy of an entity recognizer.

$sel:entityTypes:EntityRecognizerMetadata', entityRecognizerMetadata_entityTypes - Entity types from the metadata of an entity recognizer.

entityRecognizerMetadata_numberOfTestDocuments :: Lens' EntityRecognizerMetadata (Maybe Int)

The number of documents in the input data that were used to test the entity recognizer. Typically this is 10 to 20 percent of the input documents.

entityRecognizerMetadata_numberOfTrainedDocuments :: Lens' EntityRecognizerMetadata (Maybe Int)

The number of documents in the input data that were used to train the entity recognizer. Typically this is 80 to 90 percent of the input documents.

entityRecognizerMetadata_evaluationMetrics :: Lens' EntityRecognizerMetadata (Maybe EntityRecognizerEvaluationMetrics)

Detailed information about the accuracy of an entity recognizer.

entityRecognizerMetadata_entityTypes :: Lens' EntityRecognizerMetadata (Maybe [EntityRecognizerMetadataEntityTypesListItem])

Entity types from the metadata of an entity recognizer.