Network.AWS.Comprehend.Types.EntityRecognizerMetadataEntityTypesListItem

Description

Documentation

data EntityRecognizerMetadataEntityTypesListItem

Individual item from the list of entity types in the metadata of an entity recognizer.

See: newEntityRecognizerMetadataEntityTypesListItem smart constructor.

Constructors

EntityRecognizerMetadataEntityTypesListItem'

Fields

evaluationMetrics :: Maybe EntityTypesEvaluationMetrics

Detailed information about the accuracy of the entity recognizer for a specific item on the list of entity types.

type' :: Maybe Text

Type of entity from the list of entity types in the metadata of an entity recognizer.

numberOfTrainMentions :: Maybe Int

Indicates the number of times the given entity type was seen in the training data.

newEntityRecognizerMetadataEntityTypesListItem :: EntityRecognizerMetadataEntityTypesListItem

Create a value of EntityRecognizerMetadataEntityTypesListItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:evaluationMetrics:EntityRecognizerMetadataEntityTypesListItem', entityRecognizerMetadataEntityTypesListItem_evaluationMetrics - Detailed information about the accuracy of the entity recognizer for a specific item on the list of entity types.

$sel:type':EntityRecognizerMetadataEntityTypesListItem', entityRecognizerMetadataEntityTypesListItem_type - Type of entity from the list of entity types in the metadata of an entity recognizer.

$sel:numberOfTrainMentions:EntityRecognizerMetadataEntityTypesListItem', entityRecognizerMetadataEntityTypesListItem_numberOfTrainMentions - Indicates the number of times the given entity type was seen in the training data.

entityRecognizerMetadataEntityTypesListItem_evaluationMetrics :: Lens' EntityRecognizerMetadataEntityTypesListItem (Maybe EntityTypesEvaluationMetrics)

Detailed information about the accuracy of the entity recognizer for a specific item on the list of entity types.

entityRecognizerMetadataEntityTypesListItem_type :: Lens' EntityRecognizerMetadataEntityTypesListItem (Maybe Text)

Type of entity from the list of entity types in the metadata of an entity recognizer.

entityRecognizerMetadataEntityTypesListItem_numberOfTrainMentions :: Lens' EntityRecognizerMetadataEntityTypesListItem (Maybe Int)

Indicates the number of times the given entity type was seen in the training data.