Network.AWS.Comprehend.Types.EntityRecognizerProperties

Description

Documentation

data EntityRecognizerProperties

Describes information about an entity recognizer.

See: newEntityRecognizerProperties smart constructor.

Constructors

EntityRecognizerProperties'

Fields

languageCode :: Maybe LanguageCode

The language of the input documents. All documents must be in the same language. Only English ("en") is currently supported.

vpcConfig :: Maybe VpcConfig

Configuration parameters for a private Virtual Private Cloud (VPC) containing the resources you are using for your custom entity recognizer. For more information, see Amazon VPC.

status :: Maybe ModelStatus

Provides the status of the entity recognizer.

inputDataConfig :: Maybe EntityRecognizerInputDataConfig

The input data properties of an entity recognizer.

message :: Maybe Text

A description of the status of the recognizer.

versionName :: Maybe Text

The version name you assigned to the entity recognizer.

endTime :: Maybe POSIX

The time that the recognizer creation completed.

volumeKmsKeyId :: Maybe Text

ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt data on the storage volume attached to the ML compute instance(s) that process the analysis job. The VolumeKmsKeyId can be either of the following formats:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

submitTime :: Maybe POSIX

The time that the recognizer was submitted for processing.

recognizerMetadata :: Maybe (Sensitive EntityRecognizerMetadata)

Provides information about an entity recognizer.

entityRecognizerArn :: Maybe Text

The Amazon Resource Name (ARN) that identifies the entity recognizer.

modelKmsKeyId :: Maybe Text

ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt trained custom models. The ModelKmsKeyId can be either of the following formats:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

trainingStartTime :: Maybe POSIX

The time that training of the entity recognizer started.

dataAccessRoleArn :: Maybe Text

The Amazon Resource Name (ARN) of the AWS Identity and Management (IAM) role that grants Amazon Comprehend read access to your input data.

trainingEndTime :: Maybe POSIX

The time that training of the entity recognizer was completed.

newEntityRecognizerProperties :: EntityRecognizerProperties

Create a value of EntityRecognizerProperties with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:languageCode:EntityRecognizerProperties', entityRecognizerProperties_languageCode - The language of the input documents. All documents must be in the same language. Only English ("en") is currently supported.

$sel:vpcConfig:EntityRecognizerProperties', entityRecognizerProperties_vpcConfig - Configuration parameters for a private Virtual Private Cloud (VPC) containing the resources you are using for your custom entity recognizer. For more information, see Amazon VPC.

$sel:status:EntityRecognizerProperties', entityRecognizerProperties_status - Provides the status of the entity recognizer.

$sel:inputDataConfig:EntityRecognizerProperties', entityRecognizerProperties_inputDataConfig - The input data properties of an entity recognizer.

$sel:message:EntityRecognizerProperties', entityRecognizerProperties_message - A description of the status of the recognizer.

$sel:versionName:EntityRecognizerProperties', entityRecognizerProperties_versionName - The version name you assigned to the entity recognizer.

$sel:endTime:EntityRecognizerProperties', entityRecognizerProperties_endTime - The time that the recognizer creation completed.

$sel:volumeKmsKeyId:EntityRecognizerProperties', entityRecognizerProperties_volumeKmsKeyId - ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt data on the storage volume attached to the ML compute instance(s) that process the analysis job. The VolumeKmsKeyId can be either of the following formats:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

$sel:submitTime:EntityRecognizerProperties', entityRecognizerProperties_submitTime - The time that the recognizer was submitted for processing.

$sel:recognizerMetadata:EntityRecognizerProperties', entityRecognizerProperties_recognizerMetadata - Provides information about an entity recognizer.

$sel:entityRecognizerArn:EntityRecognizerProperties', entityRecognizerProperties_entityRecognizerArn - The Amazon Resource Name (ARN) that identifies the entity recognizer.

$sel:modelKmsKeyId:EntityRecognizerProperties', entityRecognizerProperties_modelKmsKeyId - ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt trained custom models. The ModelKmsKeyId can be either of the following formats:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

$sel:trainingStartTime:EntityRecognizerProperties', entityRecognizerProperties_trainingStartTime - The time that training of the entity recognizer started.

$sel:dataAccessRoleArn:EntityRecognizerProperties', entityRecognizerProperties_dataAccessRoleArn - The Amazon Resource Name (ARN) of the AWS Identity and Management (IAM) role that grants Amazon Comprehend read access to your input data.

$sel:trainingEndTime:EntityRecognizerProperties', entityRecognizerProperties_trainingEndTime - The time that training of the entity recognizer was completed.

entityRecognizerProperties_languageCode :: Lens' EntityRecognizerProperties (Maybe LanguageCode)

The language of the input documents. All documents must be in the same language. Only English ("en") is currently supported.

entityRecognizerProperties_vpcConfig :: Lens' EntityRecognizerProperties (Maybe VpcConfig)

Configuration parameters for a private Virtual Private Cloud (VPC) containing the resources you are using for your custom entity recognizer. For more information, see Amazon VPC.

entityRecognizerProperties_status :: Lens' EntityRecognizerProperties (Maybe ModelStatus)

Provides the status of the entity recognizer.

entityRecognizerProperties_inputDataConfig :: Lens' EntityRecognizerProperties (Maybe EntityRecognizerInputDataConfig)

The input data properties of an entity recognizer.

entityRecognizerProperties_message :: Lens' EntityRecognizerProperties (Maybe Text)

A description of the status of the recognizer.

entityRecognizerProperties_versionName :: Lens' EntityRecognizerProperties (Maybe Text)

The version name you assigned to the entity recognizer.

entityRecognizerProperties_endTime :: Lens' EntityRecognizerProperties (Maybe UTCTime)

The time that the recognizer creation completed.

entityRecognizerProperties_volumeKmsKeyId :: Lens' EntityRecognizerProperties (Maybe Text)

ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt data on the storage volume attached to the ML compute instance(s) that process the analysis job. The VolumeKmsKeyId can be either of the following formats:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

entityRecognizerProperties_submitTime :: Lens' EntityRecognizerProperties (Maybe UTCTime)

The time that the recognizer was submitted for processing.

entityRecognizerProperties_recognizerMetadata :: Lens' EntityRecognizerProperties (Maybe EntityRecognizerMetadata)

Provides information about an entity recognizer.

entityRecognizerProperties_entityRecognizerArn :: Lens' EntityRecognizerProperties (Maybe Text)

The Amazon Resource Name (ARN) that identifies the entity recognizer.

entityRecognizerProperties_modelKmsKeyId :: Lens' EntityRecognizerProperties (Maybe Text)

ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt trained custom models. The ModelKmsKeyId can be either of the following formats:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

entityRecognizerProperties_trainingStartTime :: Lens' EntityRecognizerProperties (Maybe UTCTime)

The time that training of the entity recognizer started.

entityRecognizerProperties_dataAccessRoleArn :: Lens' EntityRecognizerProperties (Maybe Text)

The Amazon Resource Name (ARN) of the AWS Identity and Management (IAM) role that grants Amazon Comprehend read access to your input data.

entityRecognizerProperties_trainingEndTime :: Lens' EntityRecognizerProperties (Maybe UTCTime)

The time that training of the entity recognizer was completed.