Network.AWS.Comprehend.Types.EventsDetectionJobProperties

Description

Documentation

data EventsDetectionJobProperties

Provides information about an events detection job.

See: newEventsDetectionJobProperties smart constructor.

Constructors

EventsDetectionJobProperties'

Fields

languageCode :: Maybe LanguageCode

The language code of the input documents.

jobId :: Maybe Text

The identifier assigned to the events detection job.

jobArn :: Maybe Text

The Amazon Resource Name (ARN) of the events detection job. It is a unique, fully qualified identifier for the job. It includes the AWS account, Region, and the job ID. The format of the ARN is as follows:

arn:<partition>:comprehend:<region>:<account-id>:events-detection-job/<job-id>

The following is an example job ARN:

arn:aws:comprehend:us-west-2:111122223333:events-detection-job/1234abcd12ab34cd56ef1234567890ab

jobName :: Maybe Text

The name you assigned the events detection job.

targetEventTypes :: Maybe (NonEmpty Text)

The types of events that are detected by the job.

inputDataConfig :: Maybe InputDataConfig

The input data configuration that you supplied when you created the events detection job.

endTime :: Maybe POSIX

The time that the events detection job completed.

outputDataConfig :: Maybe OutputDataConfig

The output data configuration that you supplied when you created the events detection job.

dataAccessRoleArn :: Maybe Text

The Amazon Resource Name (ARN) of the AWS Identify and Access Management (IAM) role that grants Amazon Comprehend read access to your input data.

jobStatus :: Maybe JobStatus

The current status of the events detection job.

message :: Maybe Text

A description of the status of the events detection job.

submitTime :: Maybe POSIX

The time that the events detection job was submitted for processing.

newEventsDetectionJobProperties :: EventsDetectionJobProperties

Create a value of EventsDetectionJobProperties with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:languageCode:EventsDetectionJobProperties', eventsDetectionJobProperties_languageCode - The language code of the input documents.

$sel:jobId:EventsDetectionJobProperties', eventsDetectionJobProperties_jobId - The identifier assigned to the events detection job.

$sel:jobArn:EventsDetectionJobProperties', eventsDetectionJobProperties_jobArn - The Amazon Resource Name (ARN) of the events detection job. It is a unique, fully qualified identifier for the job. It includes the AWS account, Region, and the job ID. The format of the ARN is as follows:

arn:<partition>:comprehend:<region>:<account-id>:events-detection-job/<job-id>

The following is an example job ARN:

arn:aws:comprehend:us-west-2:111122223333:events-detection-job/1234abcd12ab34cd56ef1234567890ab

$sel:jobName:EventsDetectionJobProperties', eventsDetectionJobProperties_jobName - The name you assigned the events detection job.

$sel:targetEventTypes:EventsDetectionJobProperties', eventsDetectionJobProperties_targetEventTypes - The types of events that are detected by the job.

$sel:inputDataConfig:EventsDetectionJobProperties', eventsDetectionJobProperties_inputDataConfig - The input data configuration that you supplied when you created the events detection job.

$sel:endTime:EventsDetectionJobProperties', eventsDetectionJobProperties_endTime - The time that the events detection job completed.

$sel:outputDataConfig:EventsDetectionJobProperties', eventsDetectionJobProperties_outputDataConfig - The output data configuration that you supplied when you created the events detection job.

$sel:dataAccessRoleArn:EventsDetectionJobProperties', eventsDetectionJobProperties_dataAccessRoleArn - The Amazon Resource Name (ARN) of the AWS Identify and Access Management (IAM) role that grants Amazon Comprehend read access to your input data.

$sel:jobStatus:EventsDetectionJobProperties', eventsDetectionJobProperties_jobStatus - The current status of the events detection job.

$sel:message:EventsDetectionJobProperties', eventsDetectionJobProperties_message - A description of the status of the events detection job.

$sel:submitTime:EventsDetectionJobProperties', eventsDetectionJobProperties_submitTime - The time that the events detection job was submitted for processing.

eventsDetectionJobProperties_languageCode :: Lens' EventsDetectionJobProperties (Maybe LanguageCode)

The language code of the input documents.

eventsDetectionJobProperties_jobId :: Lens' EventsDetectionJobProperties (Maybe Text)

The identifier assigned to the events detection job.

eventsDetectionJobProperties_jobArn :: Lens' EventsDetectionJobProperties (Maybe Text)

The Amazon Resource Name (ARN) of the events detection job. It is a unique, fully qualified identifier for the job. It includes the AWS account, Region, and the job ID. The format of the ARN is as follows:

arn:<partition>:comprehend:<region>:<account-id>:events-detection-job/<job-id>

The following is an example job ARN:

arn:aws:comprehend:us-west-2:111122223333:events-detection-job/1234abcd12ab34cd56ef1234567890ab

eventsDetectionJobProperties_jobName :: Lens' EventsDetectionJobProperties (Maybe Text)

The name you assigned the events detection job.

eventsDetectionJobProperties_targetEventTypes :: Lens' EventsDetectionJobProperties (Maybe (NonEmpty Text))

The types of events that are detected by the job.

eventsDetectionJobProperties_inputDataConfig :: Lens' EventsDetectionJobProperties (Maybe InputDataConfig)

The input data configuration that you supplied when you created the events detection job.

eventsDetectionJobProperties_endTime :: Lens' EventsDetectionJobProperties (Maybe UTCTime)

The time that the events detection job completed.

eventsDetectionJobProperties_outputDataConfig :: Lens' EventsDetectionJobProperties (Maybe OutputDataConfig)

The output data configuration that you supplied when you created the events detection job.

eventsDetectionJobProperties_dataAccessRoleArn :: Lens' EventsDetectionJobProperties (Maybe Text)

The Amazon Resource Name (ARN) of the AWS Identify and Access Management (IAM) role that grants Amazon Comprehend read access to your input data.

eventsDetectionJobProperties_jobStatus :: Lens' EventsDetectionJobProperties (Maybe JobStatus)

The current status of the events detection job.

eventsDetectionJobProperties_message :: Lens' EventsDetectionJobProperties (Maybe Text)

A description of the status of the events detection job.

eventsDetectionJobProperties_submitTime :: Lens' EventsDetectionJobProperties (Maybe UTCTime)

The time that the events detection job was submitted for processing.