Network.AWS.Comprehend.Types.KeyPhrasesDetectionJobProperties

Description

Documentation

data KeyPhrasesDetectionJobProperties

Provides information about a key phrases detection job.

See: newKeyPhrasesDetectionJobProperties smart constructor.

Constructors

KeyPhrasesDetectionJobProperties'

Fields

languageCode :: Maybe LanguageCode

The language code of the input documents.

vpcConfig :: Maybe VpcConfig

Configuration parameters for a private Virtual Private Cloud (VPC) containing the resources you are using for your key phrases detection job. For more information, see Amazon VPC.

inputDataConfig :: Maybe InputDataConfig

The input data configuration that you supplied when you created the key phrases detection job.

message :: Maybe Text

A description of the status of a job.

jobStatus :: Maybe JobStatus

The current status of the key phrases detection job. If the status is FAILED, the Message field shows the reason for the failure.

outputDataConfig :: Maybe OutputDataConfig

The output data configuration that you supplied when you created the key phrases detection job.

endTime :: Maybe POSIX

The time that the key phrases detection job completed.

jobArn :: Maybe Text

The Amazon Resource Name (ARN) of the key phrases detection job. It is a unique, fully qualified identifier for the job. It includes the AWS account, Region, and the job ID. The format of the ARN is as follows:

arn:<partition>:comprehend:<region>:<account-id>:key-phrases-detection-job/<job-id>

The following is an example job ARN:

arn:aws:comprehend:us-west-2:111122223333:key-phrases-detection-job/1234abcd12ab34cd56ef1234567890ab

volumeKmsKeyId :: Maybe Text

ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt data on the storage volume attached to the ML compute instance(s) that process the analysis job. The VolumeKmsKeyId can be either of the following formats:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

submitTime :: Maybe POSIX

The time that the key phrases detection job was submitted for processing.

jobName :: Maybe Text

The name that you assigned the key phrases detection job.

dataAccessRoleArn :: Maybe Text

The Amazon Resource Name (ARN) that gives Amazon Comprehend read access to your input data.

jobId :: Maybe Text

The identifier assigned to the key phrases detection job.

newKeyPhrasesDetectionJobProperties :: KeyPhrasesDetectionJobProperties

Create a value of KeyPhrasesDetectionJobProperties with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:languageCode:KeyPhrasesDetectionJobProperties', keyPhrasesDetectionJobProperties_languageCode - The language code of the input documents.

$sel:vpcConfig:KeyPhrasesDetectionJobProperties', keyPhrasesDetectionJobProperties_vpcConfig - Configuration parameters for a private Virtual Private Cloud (VPC) containing the resources you are using for your key phrases detection job. For more information, see Amazon VPC.

$sel:inputDataConfig:KeyPhrasesDetectionJobProperties', keyPhrasesDetectionJobProperties_inputDataConfig - The input data configuration that you supplied when you created the key phrases detection job.

$sel:message:KeyPhrasesDetectionJobProperties', keyPhrasesDetectionJobProperties_message - A description of the status of a job.

$sel:jobStatus:KeyPhrasesDetectionJobProperties', keyPhrasesDetectionJobProperties_jobStatus - The current status of the key phrases detection job. If the status is FAILED, the Message field shows the reason for the failure.

$sel:outputDataConfig:KeyPhrasesDetectionJobProperties', keyPhrasesDetectionJobProperties_outputDataConfig - The output data configuration that you supplied when you created the key phrases detection job.

$sel:endTime:KeyPhrasesDetectionJobProperties', keyPhrasesDetectionJobProperties_endTime - The time that the key phrases detection job completed.

$sel:jobArn:KeyPhrasesDetectionJobProperties', keyPhrasesDetectionJobProperties_jobArn - The Amazon Resource Name (ARN) of the key phrases detection job. It is a unique, fully qualified identifier for the job. It includes the AWS account, Region, and the job ID. The format of the ARN is as follows:

arn:<partition>:comprehend:<region>:<account-id>:key-phrases-detection-job/<job-id>

The following is an example job ARN:

arn:aws:comprehend:us-west-2:111122223333:key-phrases-detection-job/1234abcd12ab34cd56ef1234567890ab

$sel:volumeKmsKeyId:KeyPhrasesDetectionJobProperties', keyPhrasesDetectionJobProperties_volumeKmsKeyId - ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt data on the storage volume attached to the ML compute instance(s) that process the analysis job. The VolumeKmsKeyId can be either of the following formats:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

$sel:submitTime:KeyPhrasesDetectionJobProperties', keyPhrasesDetectionJobProperties_submitTime - The time that the key phrases detection job was submitted for processing.

$sel:jobName:KeyPhrasesDetectionJobProperties', keyPhrasesDetectionJobProperties_jobName - The name that you assigned the key phrases detection job.

$sel:dataAccessRoleArn:KeyPhrasesDetectionJobProperties', keyPhrasesDetectionJobProperties_dataAccessRoleArn - The Amazon Resource Name (ARN) that gives Amazon Comprehend read access to your input data.

$sel:jobId:KeyPhrasesDetectionJobProperties', keyPhrasesDetectionJobProperties_jobId - The identifier assigned to the key phrases detection job.

keyPhrasesDetectionJobProperties_languageCode :: Lens' KeyPhrasesDetectionJobProperties (Maybe LanguageCode)

The language code of the input documents.

keyPhrasesDetectionJobProperties_vpcConfig :: Lens' KeyPhrasesDetectionJobProperties (Maybe VpcConfig)

Configuration parameters for a private Virtual Private Cloud (VPC) containing the resources you are using for your key phrases detection job. For more information, see Amazon VPC.

keyPhrasesDetectionJobProperties_inputDataConfig :: Lens' KeyPhrasesDetectionJobProperties (Maybe InputDataConfig)

The input data configuration that you supplied when you created the key phrases detection job.

keyPhrasesDetectionJobProperties_message :: Lens' KeyPhrasesDetectionJobProperties (Maybe Text)

A description of the status of a job.

keyPhrasesDetectionJobProperties_jobStatus :: Lens' KeyPhrasesDetectionJobProperties (Maybe JobStatus)

The current status of the key phrases detection job. If the status is FAILED, the Message field shows the reason for the failure.

keyPhrasesDetectionJobProperties_outputDataConfig :: Lens' KeyPhrasesDetectionJobProperties (Maybe OutputDataConfig)

The output data configuration that you supplied when you created the key phrases detection job.

keyPhrasesDetectionJobProperties_endTime :: Lens' KeyPhrasesDetectionJobProperties (Maybe UTCTime)

The time that the key phrases detection job completed.

keyPhrasesDetectionJobProperties_jobArn :: Lens' KeyPhrasesDetectionJobProperties (Maybe Text)

The Amazon Resource Name (ARN) of the key phrases detection job. It is a unique, fully qualified identifier for the job. It includes the AWS account, Region, and the job ID. The format of the ARN is as follows:

arn:<partition>:comprehend:<region>:<account-id>:key-phrases-detection-job/<job-id>

The following is an example job ARN:

arn:aws:comprehend:us-west-2:111122223333:key-phrases-detection-job/1234abcd12ab34cd56ef1234567890ab

keyPhrasesDetectionJobProperties_volumeKmsKeyId :: Lens' KeyPhrasesDetectionJobProperties (Maybe Text)

ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt data on the storage volume attached to the ML compute instance(s) that process the analysis job. The VolumeKmsKeyId can be either of the following formats:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

keyPhrasesDetectionJobProperties_submitTime :: Lens' KeyPhrasesDetectionJobProperties (Maybe UTCTime)

The time that the key phrases detection job was submitted for processing.

keyPhrasesDetectionJobProperties_jobName :: Lens' KeyPhrasesDetectionJobProperties (Maybe Text)

The name that you assigned the key phrases detection job.

keyPhrasesDetectionJobProperties_dataAccessRoleArn :: Lens' KeyPhrasesDetectionJobProperties (Maybe Text)

The Amazon Resource Name (ARN) that gives Amazon Comprehend read access to your input data.

keyPhrasesDetectionJobProperties_jobId :: Lens' KeyPhrasesDetectionJobProperties (Maybe Text)

The identifier assigned to the key phrases detection job.