Network.AWS.Comprehend.Types.PiiEntitiesDetectionJobFilter

Description

Documentation

data PiiEntitiesDetectionJobFilter

Provides information for filtering a list of PII entity detection jobs.

See: newPiiEntitiesDetectionJobFilter smart constructor.

Constructors

PiiEntitiesDetectionJobFilter'

Fields

jobStatus :: Maybe JobStatus

Filters the list of jobs based on job status. Returns only jobs with the specified status.

submitTimeBefore :: Maybe POSIX

Filters the list of jobs based on the time that the job was submitted for processing. Returns only jobs submitted before the specified time. Jobs are returned in ascending order, oldest to newest.

submitTimeAfter :: Maybe POSIX

Filters the list of jobs based on the time that the job was submitted for processing. Returns only jobs submitted after the specified time. Jobs are returned in descending order, newest to oldest.

jobName :: Maybe Text

Filters on the name of the job.

newPiiEntitiesDetectionJobFilter :: PiiEntitiesDetectionJobFilter

Create a value of PiiEntitiesDetectionJobFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobStatus:PiiEntitiesDetectionJobFilter', piiEntitiesDetectionJobFilter_jobStatus - Filters the list of jobs based on job status. Returns only jobs with the specified status.

$sel:submitTimeBefore:PiiEntitiesDetectionJobFilter', piiEntitiesDetectionJobFilter_submitTimeBefore - Filters the list of jobs based on the time that the job was submitted for processing. Returns only jobs submitted before the specified time. Jobs are returned in ascending order, oldest to newest.

$sel:submitTimeAfter:PiiEntitiesDetectionJobFilter', piiEntitiesDetectionJobFilter_submitTimeAfter - Filters the list of jobs based on the time that the job was submitted for processing. Returns only jobs submitted after the specified time. Jobs are returned in descending order, newest to oldest.

$sel:jobName:PiiEntitiesDetectionJobFilter', piiEntitiesDetectionJobFilter_jobName - Filters on the name of the job.

piiEntitiesDetectionJobFilter_jobStatus :: Lens' PiiEntitiesDetectionJobFilter (Maybe JobStatus)

Filters the list of jobs based on job status. Returns only jobs with the specified status.

piiEntitiesDetectionJobFilter_submitTimeBefore :: Lens' PiiEntitiesDetectionJobFilter (Maybe UTCTime)

Filters the list of jobs based on the time that the job was submitted for processing. Returns only jobs submitted before the specified time. Jobs are returned in ascending order, oldest to newest.

piiEntitiesDetectionJobFilter_submitTimeAfter :: Lens' PiiEntitiesDetectionJobFilter (Maybe UTCTime)

Filters the list of jobs based on the time that the job was submitted for processing. Returns only jobs submitted after the specified time. Jobs are returned in descending order, newest to oldest.

piiEntitiesDetectionJobFilter_jobName :: Lens' PiiEntitiesDetectionJobFilter (Maybe Text)

Filters on the name of the job.