Network.AWS.Comprehend.Types.PiiEntitiesDetectionJobProperties

Description

Documentation

data PiiEntitiesDetectionJobProperties

Provides information about a PII entities detection job.

See: newPiiEntitiesDetectionJobProperties smart constructor.

Constructors

PiiEntitiesDetectionJobProperties'

Fields

languageCode :: Maybe LanguageCode

The language code of the input documents

redactionConfig :: Maybe RedactionConfig

Provides configuration parameters for PII entity redaction.

This parameter is required if you set the Mode parameter to ONLY_REDACTION. In that case, you must provide a RedactionConfig definition that includes the PiiEntityTypes parameter.

inputDataConfig :: Maybe InputDataConfig

The input properties for a PII entities detection job.

message :: Maybe Text

A description of the status of a job.

mode :: Maybe PiiEntitiesDetectionMode

Specifies whether the output provides the locations (offsets) of PII entities or a file in which PII entities are redacted.

jobStatus :: Maybe JobStatus

The current status of the PII entities detection job. If the status is FAILED, the Message field shows the reason for the failure.

outputDataConfig :: Maybe PiiOutputDataConfig

The output data configuration that you supplied when you created the PII entities detection job.

endTime :: Maybe POSIX

The time that the PII entities detection job completed.

jobArn :: Maybe Text

The Amazon Resource Name (ARN) of the PII entities detection job. It is a unique, fully qualified identifier for the job. It includes the AWS account, Region, and the job ID. The format of the ARN is as follows:

arn:<partition>:comprehend:<region>:<account-id>:pii-entities-detection-job/<job-id>

The following is an example job ARN:

arn:aws:comprehend:us-west-2:111122223333:pii-entities-detection-job/1234abcd12ab34cd56ef1234567890ab

submitTime :: Maybe POSIX

The time that the PII entities detection job was submitted for processing.

jobName :: Maybe Text

The name that you assigned the PII entities detection job.

dataAccessRoleArn :: Maybe Text

The Amazon Resource Name (ARN) that gives Amazon Comprehend read access to your input data.

jobId :: Maybe Text

The identifier assigned to the PII entities detection job.

newPiiEntitiesDetectionJobProperties :: PiiEntitiesDetectionJobProperties

Create a value of PiiEntitiesDetectionJobProperties with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:languageCode:PiiEntitiesDetectionJobProperties', piiEntitiesDetectionJobProperties_languageCode - The language code of the input documents

$sel:redactionConfig:PiiEntitiesDetectionJobProperties', piiEntitiesDetectionJobProperties_redactionConfig - Provides configuration parameters for PII entity redaction.

This parameter is required if you set the Mode parameter to ONLY_REDACTION. In that case, you must provide a RedactionConfig definition that includes the PiiEntityTypes parameter.

$sel:inputDataConfig:PiiEntitiesDetectionJobProperties', piiEntitiesDetectionJobProperties_inputDataConfig - The input properties for a PII entities detection job.

$sel:message:PiiEntitiesDetectionJobProperties', piiEntitiesDetectionJobProperties_message - A description of the status of a job.

$sel:mode:PiiEntitiesDetectionJobProperties', piiEntitiesDetectionJobProperties_mode - Specifies whether the output provides the locations (offsets) of PII entities or a file in which PII entities are redacted.

$sel:jobStatus:PiiEntitiesDetectionJobProperties', piiEntitiesDetectionJobProperties_jobStatus - The current status of the PII entities detection job. If the status is FAILED, the Message field shows the reason for the failure.

$sel:outputDataConfig:PiiEntitiesDetectionJobProperties', piiEntitiesDetectionJobProperties_outputDataConfig - The output data configuration that you supplied when you created the PII entities detection job.

$sel:endTime:PiiEntitiesDetectionJobProperties', piiEntitiesDetectionJobProperties_endTime - The time that the PII entities detection job completed.

$sel:jobArn:PiiEntitiesDetectionJobProperties', piiEntitiesDetectionJobProperties_jobArn - The Amazon Resource Name (ARN) of the PII entities detection job. It is a unique, fully qualified identifier for the job. It includes the AWS account, Region, and the job ID. The format of the ARN is as follows:

arn:<partition>:comprehend:<region>:<account-id>:pii-entities-detection-job/<job-id>

The following is an example job ARN:

arn:aws:comprehend:us-west-2:111122223333:pii-entities-detection-job/1234abcd12ab34cd56ef1234567890ab

$sel:submitTime:PiiEntitiesDetectionJobProperties', piiEntitiesDetectionJobProperties_submitTime - The time that the PII entities detection job was submitted for processing.

$sel:jobName:PiiEntitiesDetectionJobProperties', piiEntitiesDetectionJobProperties_jobName - The name that you assigned the PII entities detection job.

$sel:dataAccessRoleArn:PiiEntitiesDetectionJobProperties', piiEntitiesDetectionJobProperties_dataAccessRoleArn - The Amazon Resource Name (ARN) that gives Amazon Comprehend read access to your input data.

$sel:jobId:PiiEntitiesDetectionJobProperties', piiEntitiesDetectionJobProperties_jobId - The identifier assigned to the PII entities detection job.

piiEntitiesDetectionJobProperties_languageCode :: Lens' PiiEntitiesDetectionJobProperties (Maybe LanguageCode)

The language code of the input documents

piiEntitiesDetectionJobProperties_redactionConfig :: Lens' PiiEntitiesDetectionJobProperties (Maybe RedactionConfig)

Provides configuration parameters for PII entity redaction.

This parameter is required if you set the Mode parameter to ONLY_REDACTION. In that case, you must provide a RedactionConfig definition that includes the PiiEntityTypes parameter.

piiEntitiesDetectionJobProperties_inputDataConfig :: Lens' PiiEntitiesDetectionJobProperties (Maybe InputDataConfig)

The input properties for a PII entities detection job.

piiEntitiesDetectionJobProperties_message :: Lens' PiiEntitiesDetectionJobProperties (Maybe Text)

A description of the status of a job.

piiEntitiesDetectionJobProperties_mode :: Lens' PiiEntitiesDetectionJobProperties (Maybe PiiEntitiesDetectionMode)

Specifies whether the output provides the locations (offsets) of PII entities or a file in which PII entities are redacted.

piiEntitiesDetectionJobProperties_jobStatus :: Lens' PiiEntitiesDetectionJobProperties (Maybe JobStatus)

The current status of the PII entities detection job. If the status is FAILED, the Message field shows the reason for the failure.

piiEntitiesDetectionJobProperties_outputDataConfig :: Lens' PiiEntitiesDetectionJobProperties (Maybe PiiOutputDataConfig)

The output data configuration that you supplied when you created the PII entities detection job.

piiEntitiesDetectionJobProperties_endTime :: Lens' PiiEntitiesDetectionJobProperties (Maybe UTCTime)

The time that the PII entities detection job completed.

piiEntitiesDetectionJobProperties_jobArn :: Lens' PiiEntitiesDetectionJobProperties (Maybe Text)

The Amazon Resource Name (ARN) of the PII entities detection job. It is a unique, fully qualified identifier for the job. It includes the AWS account, Region, and the job ID. The format of the ARN is as follows:

arn:<partition>:comprehend:<region>:<account-id>:pii-entities-detection-job/<job-id>

The following is an example job ARN:

arn:aws:comprehend:us-west-2:111122223333:pii-entities-detection-job/1234abcd12ab34cd56ef1234567890ab

piiEntitiesDetectionJobProperties_submitTime :: Lens' PiiEntitiesDetectionJobProperties (Maybe UTCTime)

The time that the PII entities detection job was submitted for processing.

piiEntitiesDetectionJobProperties_jobName :: Lens' PiiEntitiesDetectionJobProperties (Maybe Text)

The name that you assigned the PII entities detection job.

piiEntitiesDetectionJobProperties_dataAccessRoleArn :: Lens' PiiEntitiesDetectionJobProperties (Maybe Text)

The Amazon Resource Name (ARN) that gives Amazon Comprehend read access to your input data.

piiEntitiesDetectionJobProperties_jobId :: Lens' PiiEntitiesDetectionJobProperties (Maybe Text)

The identifier assigned to the PII entities detection job.