Network.AWS.Comprehend.Types.SentimentDetectionJobFilter

Description

Documentation

data SentimentDetectionJobFilter

Provides information for filtering a list of dominant language detection jobs. For more information, see the operation.

See: newSentimentDetectionJobFilter smart constructor.

Constructors

SentimentDetectionJobFilter'

Fields

submitTimeAfter :: Maybe POSIX

Filters the list of jobs based on the time that the job was submitted for processing. Returns only jobs submitted after the specified time. Jobs are returned in descending order, newest to oldest.

submitTimeBefore :: Maybe POSIX

Filters the list of jobs based on the time that the job was submitted for processing. Returns only jobs submitted before the specified time. Jobs are returned in ascending order, oldest to newest.

jobName :: Maybe Text

Filters on the name of the job.

jobStatus :: Maybe JobStatus

Filters the list of jobs based on job status. Returns only jobs with the specified status.

newSentimentDetectionJobFilter :: SentimentDetectionJobFilter

Create a value of SentimentDetectionJobFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:submitTimeAfter:SentimentDetectionJobFilter', sentimentDetectionJobFilter_submitTimeAfter - Filters the list of jobs based on the time that the job was submitted for processing. Returns only jobs submitted after the specified time. Jobs are returned in descending order, newest to oldest.

$sel:submitTimeBefore:SentimentDetectionJobFilter', sentimentDetectionJobFilter_submitTimeBefore - Filters the list of jobs based on the time that the job was submitted for processing. Returns only jobs submitted before the specified time. Jobs are returned in ascending order, oldest to newest.

$sel:jobName:SentimentDetectionJobFilter', sentimentDetectionJobFilter_jobName - Filters on the name of the job.

$sel:jobStatus:SentimentDetectionJobFilter', sentimentDetectionJobFilter_jobStatus - Filters the list of jobs based on job status. Returns only jobs with the specified status.

sentimentDetectionJobFilter_submitTimeAfter :: Lens' SentimentDetectionJobFilter (Maybe UTCTime)

Filters the list of jobs based on the time that the job was submitted for processing. Returns only jobs submitted after the specified time. Jobs are returned in descending order, newest to oldest.

sentimentDetectionJobFilter_submitTimeBefore :: Lens' SentimentDetectionJobFilter (Maybe UTCTime)

Filters the list of jobs based on the time that the job was submitted for processing. Returns only jobs submitted before the specified time. Jobs are returned in ascending order, oldest to newest.

sentimentDetectionJobFilter_jobName :: Lens' SentimentDetectionJobFilter (Maybe Text)

Filters on the name of the job.

sentimentDetectionJobFilter_jobStatus :: Lens' SentimentDetectionJobFilter (Maybe JobStatus)

Filters the list of jobs based on job status. Returns only jobs with the specified status.