Network.AWS.Comprehend.Types.SentimentDetectionJobProperties

Description

Documentation

data SentimentDetectionJobProperties

Provides information about a sentiment detection job.

See: newSentimentDetectionJobProperties smart constructor.

Constructors

SentimentDetectionJobProperties'

Fields

languageCode :: Maybe LanguageCode

The language code of the input documents.

jobId :: Maybe Text

The identifier assigned to the sentiment detection job.

jobArn :: Maybe Text

The Amazon Resource Name (ARN) of the sentiment detection job. It is a unique, fully qualified identifier for the job. It includes the AWS account, Region, and the job ID. The format of the ARN is as follows:

arn:<partition>:comprehend:<region>:<account-id>:sentiment-detection-job/<job-id>

The following is an example job ARN:

arn:aws:comprehend:us-west-2:111122223333:sentiment-detection-job/1234abcd12ab34cd56ef1234567890ab

jobName :: Maybe Text

The name that you assigned to the sentiment detection job

inputDataConfig :: Maybe InputDataConfig

The input data configuration that you supplied when you created the sentiment detection job.

vpcConfig :: Maybe VpcConfig

Configuration parameters for a private Virtual Private Cloud (VPC) containing the resources you are using for your sentiment detection job. For more information, see Amazon VPC.

volumeKmsKeyId :: Maybe Text

ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt data on the storage volume attached to the ML compute instance(s) that process the analysis job. The VolumeKmsKeyId can be either of the following formats:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

endTime :: Maybe POSIX

The time that the sentiment detection job ended.

outputDataConfig :: Maybe OutputDataConfig

The output data configuration that you supplied when you created the sentiment detection job.

dataAccessRoleArn :: Maybe Text

The Amazon Resource Name (ARN) that gives Amazon Comprehend read access to your input data.

jobStatus :: Maybe JobStatus

The current status of the sentiment detection job. If the status is FAILED, the Messages field shows the reason for the failure.

message :: Maybe Text

A description of the status of a job.

submitTime :: Maybe POSIX

The time that the sentiment detection job was submitted for processing.

newSentimentDetectionJobProperties :: SentimentDetectionJobProperties

Create a value of SentimentDetectionJobProperties with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:languageCode:SentimentDetectionJobProperties', sentimentDetectionJobProperties_languageCode - The language code of the input documents.

$sel:jobId:SentimentDetectionJobProperties', sentimentDetectionJobProperties_jobId - The identifier assigned to the sentiment detection job.

$sel:jobArn:SentimentDetectionJobProperties', sentimentDetectionJobProperties_jobArn - The Amazon Resource Name (ARN) of the sentiment detection job. It is a unique, fully qualified identifier for the job. It includes the AWS account, Region, and the job ID. The format of the ARN is as follows:

arn:<partition>:comprehend:<region>:<account-id>:sentiment-detection-job/<job-id>

The following is an example job ARN:

arn:aws:comprehend:us-west-2:111122223333:sentiment-detection-job/1234abcd12ab34cd56ef1234567890ab

$sel:jobName:SentimentDetectionJobProperties', sentimentDetectionJobProperties_jobName - The name that you assigned to the sentiment detection job

$sel:inputDataConfig:SentimentDetectionJobProperties', sentimentDetectionJobProperties_inputDataConfig - The input data configuration that you supplied when you created the sentiment detection job.

$sel:vpcConfig:SentimentDetectionJobProperties', sentimentDetectionJobProperties_vpcConfig - Configuration parameters for a private Virtual Private Cloud (VPC) containing the resources you are using for your sentiment detection job. For more information, see Amazon VPC.

$sel:volumeKmsKeyId:SentimentDetectionJobProperties', sentimentDetectionJobProperties_volumeKmsKeyId - ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt data on the storage volume attached to the ML compute instance(s) that process the analysis job. The VolumeKmsKeyId can be either of the following formats:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

$sel:endTime:SentimentDetectionJobProperties', sentimentDetectionJobProperties_endTime - The time that the sentiment detection job ended.

$sel:outputDataConfig:SentimentDetectionJobProperties', sentimentDetectionJobProperties_outputDataConfig - The output data configuration that you supplied when you created the sentiment detection job.

$sel:dataAccessRoleArn:SentimentDetectionJobProperties', sentimentDetectionJobProperties_dataAccessRoleArn - The Amazon Resource Name (ARN) that gives Amazon Comprehend read access to your input data.

$sel:jobStatus:SentimentDetectionJobProperties', sentimentDetectionJobProperties_jobStatus - The current status of the sentiment detection job. If the status is FAILED, the Messages field shows the reason for the failure.

$sel:message:SentimentDetectionJobProperties', sentimentDetectionJobProperties_message - A description of the status of a job.

$sel:submitTime:SentimentDetectionJobProperties', sentimentDetectionJobProperties_submitTime - The time that the sentiment detection job was submitted for processing.

sentimentDetectionJobProperties_languageCode :: Lens' SentimentDetectionJobProperties (Maybe LanguageCode)

The language code of the input documents.

sentimentDetectionJobProperties_jobId :: Lens' SentimentDetectionJobProperties (Maybe Text)

The identifier assigned to the sentiment detection job.

sentimentDetectionJobProperties_jobArn :: Lens' SentimentDetectionJobProperties (Maybe Text)

The Amazon Resource Name (ARN) of the sentiment detection job. It is a unique, fully qualified identifier for the job. It includes the AWS account, Region, and the job ID. The format of the ARN is as follows:

arn:<partition>:comprehend:<region>:<account-id>:sentiment-detection-job/<job-id>

The following is an example job ARN:

arn:aws:comprehend:us-west-2:111122223333:sentiment-detection-job/1234abcd12ab34cd56ef1234567890ab

sentimentDetectionJobProperties_jobName :: Lens' SentimentDetectionJobProperties (Maybe Text)

The name that you assigned to the sentiment detection job

sentimentDetectionJobProperties_inputDataConfig :: Lens' SentimentDetectionJobProperties (Maybe InputDataConfig)

The input data configuration that you supplied when you created the sentiment detection job.

sentimentDetectionJobProperties_vpcConfig :: Lens' SentimentDetectionJobProperties (Maybe VpcConfig)

Configuration parameters for a private Virtual Private Cloud (VPC) containing the resources you are using for your sentiment detection job. For more information, see Amazon VPC.

sentimentDetectionJobProperties_volumeKmsKeyId :: Lens' SentimentDetectionJobProperties (Maybe Text)

ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt data on the storage volume attached to the ML compute instance(s) that process the analysis job. The VolumeKmsKeyId can be either of the following formats:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

sentimentDetectionJobProperties_endTime :: Lens' SentimentDetectionJobProperties (Maybe UTCTime)

The time that the sentiment detection job ended.

sentimentDetectionJobProperties_outputDataConfig :: Lens' SentimentDetectionJobProperties (Maybe OutputDataConfig)

The output data configuration that you supplied when you created the sentiment detection job.

sentimentDetectionJobProperties_dataAccessRoleArn :: Lens' SentimentDetectionJobProperties (Maybe Text)

The Amazon Resource Name (ARN) that gives Amazon Comprehend read access to your input data.

sentimentDetectionJobProperties_jobStatus :: Lens' SentimentDetectionJobProperties (Maybe JobStatus)

The current status of the sentiment detection job. If the status is FAILED, the Messages field shows the reason for the failure.

sentimentDetectionJobProperties_message :: Lens' SentimentDetectionJobProperties (Maybe Text)

A description of the status of a job.

sentimentDetectionJobProperties_submitTime :: Lens' SentimentDetectionJobProperties (Maybe UTCTime)

The time that the sentiment detection job was submitted for processing.