Network.AWS.Comprehend.Types.SentimentScore

Description

Documentation

data SentimentScore

Describes the level of confidence that Amazon Comprehend has in the accuracy of its detection of sentiments.

See: newSentimentScore smart constructor.

Constructors

SentimentScore'

Fields

negative :: Maybe Double

The level of confidence that Amazon Comprehend has in the accuracy of its detection of the NEGATIVE sentiment.

mixed :: Maybe Double

The level of confidence that Amazon Comprehend has in the accuracy of its detection of the MIXED sentiment.

positive :: Maybe Double

The level of confidence that Amazon Comprehend has in the accuracy of its detection of the POSITIVE sentiment.

neutral :: Maybe Double

The level of confidence that Amazon Comprehend has in the accuracy of its detection of the NEUTRAL sentiment.

newSentimentScore :: SentimentScore

Create a value of SentimentScore with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:negative:SentimentScore', sentimentScore_negative - The level of confidence that Amazon Comprehend has in the accuracy of its detection of the NEGATIVE sentiment.

$sel:mixed:SentimentScore', sentimentScore_mixed - The level of confidence that Amazon Comprehend has in the accuracy of its detection of the MIXED sentiment.

$sel:positive:SentimentScore', sentimentScore_positive - The level of confidence that Amazon Comprehend has in the accuracy of its detection of the POSITIVE sentiment.

$sel:neutral:SentimentScore', sentimentScore_neutral - The level of confidence that Amazon Comprehend has in the accuracy of its detection of the NEUTRAL sentiment.

sentimentScore_negative :: Lens' SentimentScore (Maybe Double)

The level of confidence that Amazon Comprehend has in the accuracy of its detection of the NEGATIVE sentiment.

sentimentScore_mixed :: Lens' SentimentScore (Maybe Double)

The level of confidence that Amazon Comprehend has in the accuracy of its detection of the MIXED sentiment.

sentimentScore_positive :: Lens' SentimentScore (Maybe Double)

The level of confidence that Amazon Comprehend has in the accuracy of its detection of the POSITIVE sentiment.

sentimentScore_neutral :: Lens' SentimentScore (Maybe Double)

The level of confidence that Amazon Comprehend has in the accuracy of its detection of the NEUTRAL sentiment.