Network.AWS.Comprehend.Types.TopicsDetectionJobFilter

Description

Documentation

data TopicsDetectionJobFilter

Provides information for filtering topic detection jobs. For more information, see .

See: newTopicsDetectionJobFilter smart constructor.

Constructors

TopicsDetectionJobFilter'

Fields

jobStatus :: Maybe JobStatus

Filters the list of topic detection jobs based on job status. Returns only jobs with the specified status.

submitTimeBefore :: Maybe POSIX

Filters the list of jobs based on the time that the job was submitted for processing. Only returns jobs submitted before the specified time. Jobs are returned in descending order, newest to oldest.

submitTimeAfter :: Maybe POSIX

Filters the list of jobs based on the time that the job was submitted for processing. Only returns jobs submitted after the specified time. Jobs are returned in ascending order, oldest to newest.

jobName :: Maybe Text

newTopicsDetectionJobFilter :: TopicsDetectionJobFilter

Create a value of TopicsDetectionJobFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobStatus:TopicsDetectionJobFilter', topicsDetectionJobFilter_jobStatus - Filters the list of topic detection jobs based on job status. Returns only jobs with the specified status.

$sel:submitTimeBefore:TopicsDetectionJobFilter', topicsDetectionJobFilter_submitTimeBefore - Filters the list of jobs based on the time that the job was submitted for processing. Only returns jobs submitted before the specified time. Jobs are returned in descending order, newest to oldest.

$sel:submitTimeAfter:TopicsDetectionJobFilter', topicsDetectionJobFilter_submitTimeAfter - Filters the list of jobs based on the time that the job was submitted for processing. Only returns jobs submitted after the specified time. Jobs are returned in ascending order, oldest to newest.

$sel:jobName:TopicsDetectionJobFilter', topicsDetectionJobFilter_jobName -

topicsDetectionJobFilter_jobStatus :: Lens' TopicsDetectionJobFilter (Maybe JobStatus)

Filters the list of topic detection jobs based on job status. Returns only jobs with the specified status.

topicsDetectionJobFilter_submitTimeBefore :: Lens' TopicsDetectionJobFilter (Maybe UTCTime)

Filters the list of jobs based on the time that the job was submitted for processing. Only returns jobs submitted before the specified time. Jobs are returned in descending order, newest to oldest.

topicsDetectionJobFilter_submitTimeAfter :: Lens' TopicsDetectionJobFilter (Maybe UTCTime)

Filters the list of jobs based on the time that the job was submitted for processing. Only returns jobs submitted after the specified time. Jobs are returned in ascending order, oldest to newest.

topicsDetectionJobFilter_jobName :: Lens' TopicsDetectionJobFilter (Maybe Text)