Network.AWS.Comprehend.Types.TopicsDetectionJobProperties

Description

Documentation

data TopicsDetectionJobProperties

Provides information about a topic detection job.

See: newTopicsDetectionJobProperties smart constructor.

Constructors

TopicsDetectionJobProperties'

Fields

vpcConfig :: Maybe VpcConfig

Configuration parameters for a private Virtual Private Cloud (VPC) containing the resources you are using for your topic detection job. For more information, see Amazon VPC.

inputDataConfig :: Maybe InputDataConfig

The input data configuration supplied when you created the topic detection job.

message :: Maybe Text

A description for the status of a job.

jobStatus :: Maybe JobStatus

The current status of the topic detection job. If the status is Failed, the reason for the failure is shown in the Message field.

outputDataConfig :: Maybe OutputDataConfig

The output data configuration supplied when you created the topic detection job.

endTime :: Maybe POSIX

The time that the topic detection job was completed.

jobArn :: Maybe Text

The Amazon Resource Name (ARN) of the topics detection job. It is a unique, fully qualified identifier for the job. It includes the AWS account, Region, and the job ID. The format of the ARN is as follows:

arn:<partition>:comprehend:<region>:<account-id>:topics-detection-job/<job-id>

The following is an example job ARN:

arn:aws:comprehend:us-west-2:111122223333:topics-detection-job/1234abcd12ab34cd56ef1234567890ab

volumeKmsKeyId :: Maybe Text

ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt data on the storage volume attached to the ML compute instance(s) that process the analysis job. The VolumeKmsKeyId can be either of the following formats:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

submitTime :: Maybe POSIX

The time that the topic detection job was submitted for processing.

numberOfTopics :: Maybe Int

The number of topics to detect supplied when you created the topic detection job. The default is 10.

jobName :: Maybe Text

The name of the topic detection job.

dataAccessRoleArn :: Maybe Text

The Amazon Resource Name (ARN) of the AWS Identity and Management (IAM) role that grants Amazon Comprehend read access to your job data.

jobId :: Maybe Text

The identifier assigned to the topic detection job.

newTopicsDetectionJobProperties :: TopicsDetectionJobProperties

Create a value of TopicsDetectionJobProperties with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpcConfig:TopicsDetectionJobProperties', topicsDetectionJobProperties_vpcConfig - Configuration parameters for a private Virtual Private Cloud (VPC) containing the resources you are using for your topic detection job. For more information, see Amazon VPC.

$sel:inputDataConfig:TopicsDetectionJobProperties', topicsDetectionJobProperties_inputDataConfig - The input data configuration supplied when you created the topic detection job.

$sel:message:TopicsDetectionJobProperties', topicsDetectionJobProperties_message - A description for the status of a job.

$sel:jobStatus:TopicsDetectionJobProperties', topicsDetectionJobProperties_jobStatus - The current status of the topic detection job. If the status is Failed, the reason for the failure is shown in the Message field.

$sel:outputDataConfig:TopicsDetectionJobProperties', topicsDetectionJobProperties_outputDataConfig - The output data configuration supplied when you created the topic detection job.

$sel:endTime:TopicsDetectionJobProperties', topicsDetectionJobProperties_endTime - The time that the topic detection job was completed.

$sel:jobArn:TopicsDetectionJobProperties', topicsDetectionJobProperties_jobArn - The Amazon Resource Name (ARN) of the topics detection job. It is a unique, fully qualified identifier for the job. It includes the AWS account, Region, and the job ID. The format of the ARN is as follows:

arn:<partition>:comprehend:<region>:<account-id>:topics-detection-job/<job-id>

The following is an example job ARN:

arn:aws:comprehend:us-west-2:111122223333:topics-detection-job/1234abcd12ab34cd56ef1234567890ab

$sel:volumeKmsKeyId:TopicsDetectionJobProperties', topicsDetectionJobProperties_volumeKmsKeyId - ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt data on the storage volume attached to the ML compute instance(s) that process the analysis job. The VolumeKmsKeyId can be either of the following formats:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

$sel:submitTime:TopicsDetectionJobProperties', topicsDetectionJobProperties_submitTime - The time that the topic detection job was submitted for processing.

$sel:numberOfTopics:TopicsDetectionJobProperties', topicsDetectionJobProperties_numberOfTopics - The number of topics to detect supplied when you created the topic detection job. The default is 10.

$sel:jobName:TopicsDetectionJobProperties', topicsDetectionJobProperties_jobName - The name of the topic detection job.

$sel:dataAccessRoleArn:TopicsDetectionJobProperties', topicsDetectionJobProperties_dataAccessRoleArn - The Amazon Resource Name (ARN) of the AWS Identity and Management (IAM) role that grants Amazon Comprehend read access to your job data.

$sel:jobId:TopicsDetectionJobProperties', topicsDetectionJobProperties_jobId - The identifier assigned to the topic detection job.

topicsDetectionJobProperties_vpcConfig :: Lens' TopicsDetectionJobProperties (Maybe VpcConfig)

Configuration parameters for a private Virtual Private Cloud (VPC) containing the resources you are using for your topic detection job. For more information, see Amazon VPC.

topicsDetectionJobProperties_inputDataConfig :: Lens' TopicsDetectionJobProperties (Maybe InputDataConfig)

The input data configuration supplied when you created the topic detection job.

topicsDetectionJobProperties_message :: Lens' TopicsDetectionJobProperties (Maybe Text)

A description for the status of a job.

topicsDetectionJobProperties_jobStatus :: Lens' TopicsDetectionJobProperties (Maybe JobStatus)

The current status of the topic detection job. If the status is Failed, the reason for the failure is shown in the Message field.

topicsDetectionJobProperties_outputDataConfig :: Lens' TopicsDetectionJobProperties (Maybe OutputDataConfig)

The output data configuration supplied when you created the topic detection job.

topicsDetectionJobProperties_endTime :: Lens' TopicsDetectionJobProperties (Maybe UTCTime)

The time that the topic detection job was completed.

topicsDetectionJobProperties_jobArn :: Lens' TopicsDetectionJobProperties (Maybe Text)

The Amazon Resource Name (ARN) of the topics detection job. It is a unique, fully qualified identifier for the job. It includes the AWS account, Region, and the job ID. The format of the ARN is as follows:

arn:<partition>:comprehend:<region>:<account-id>:topics-detection-job/<job-id>

The following is an example job ARN:

arn:aws:comprehend:us-west-2:111122223333:topics-detection-job/1234abcd12ab34cd56ef1234567890ab

topicsDetectionJobProperties_volumeKmsKeyId :: Lens' TopicsDetectionJobProperties (Maybe Text)

ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt data on the storage volume attached to the ML compute instance(s) that process the analysis job. The VolumeKmsKeyId can be either of the following formats:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

topicsDetectionJobProperties_submitTime :: Lens' TopicsDetectionJobProperties (Maybe UTCTime)

The time that the topic detection job was submitted for processing.

topicsDetectionJobProperties_numberOfTopics :: Lens' TopicsDetectionJobProperties (Maybe Int)

The number of topics to detect supplied when you created the topic detection job. The default is 10.

topicsDetectionJobProperties_jobName :: Lens' TopicsDetectionJobProperties (Maybe Text)

The name of the topic detection job.

topicsDetectionJobProperties_dataAccessRoleArn :: Lens' TopicsDetectionJobProperties (Maybe Text)

The Amazon Resource Name (ARN) of the AWS Identity and Management (IAM) role that grants Amazon Comprehend read access to your job data.

topicsDetectionJobProperties_jobId :: Lens' TopicsDetectionJobProperties (Maybe Text)

The identifier assigned to the topic detection job.