Network.AWS.ComprehendMedical.DetectPHI

Description

Inspects the clinical text for protected health information (PHI) entities and returns the entity category, location, and confidence score for each entity. Amazon Comprehend Medical only detects entities in English language texts.

Creating a Request

data DetectPHI

See: newDetectPHI smart constructor.

Constructors

DetectPHI'

Fields

text :: Text

A UTF-8 text string containing the clinical content being examined for PHI entities. Each string must contain fewer than 20,000 bytes of characters.

newDetectPHI

Arguments

:: Text
$sel:text:DetectPHI'

-> DetectPHI

Create a value of DetectPHI with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:text:DetectPHI', detectPHI_text - A UTF-8 text string containing the clinical content being examined for PHI entities. Each string must contain fewer than 20,000 bytes of characters.

Request Lenses

detectPHI_text :: Lens' DetectPHI Text

A UTF-8 text string containing the clinical content being examined for PHI entities. Each string must contain fewer than 20,000 bytes of characters.

Destructuring the Response

data DetectPHIResponse

See: newDetectPHIResponse smart constructor.

Constructors

DetectPHIResponse'

Fields

paginationToken :: Maybe Text

If the result of the previous request to DetectPHI was truncated, include the PaginationToken to fetch the next page of PHI entities.

httpStatus :: Int

The response's http status code.

entities :: [Entity]

The collection of PHI entities extracted from the input text and their associated information. For each entity, the response provides the entity text, the entity category, where the entity text begins and ends, and the level of confidence that Amazon Comprehend Medical has in its detection.

modelVersion :: Text

The version of the model used to analyze the documents. The version number looks like X.X.X. You can use this information to track the model used for a particular batch of documents.

newDetectPHIResponse

Arguments

:: Int
$sel:httpStatus:DetectPHIResponse'

-> Text
$sel:modelVersion:DetectPHIResponse'

-> DetectPHIResponse

Create a value of DetectPHIResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:paginationToken:DetectPHIResponse', detectPHIResponse_paginationToken - If the result of the previous request to DetectPHI was truncated, include the PaginationToken to fetch the next page of PHI entities.

$sel:httpStatus:DetectPHIResponse', detectPHIResponse_httpStatus - The response's http status code.

$sel:entities:DetectPHIResponse', detectPHIResponse_entities - The collection of PHI entities extracted from the input text and their associated information. For each entity, the response provides the entity text, the entity category, where the entity text begins and ends, and the level of confidence that Amazon Comprehend Medical has in its detection.

$sel:modelVersion:DetectPHIResponse', detectPHIResponse_modelVersion - The version of the model used to analyze the documents. The version number looks like X.X.X. You can use this information to track the model used for a particular batch of documents.

Response Lenses

detectPHIResponse_paginationToken :: Lens' DetectPHIResponse (Maybe Text)

If the result of the previous request to DetectPHI was truncated, include the PaginationToken to fetch the next page of PHI entities.

detectPHIResponse_httpStatus :: Lens' DetectPHIResponse Int

The response's http status code.

detectPHIResponse_entities :: Lens' DetectPHIResponse [Entity]

The collection of PHI entities extracted from the input text and their associated information. For each entity, the response provides the entity text, the entity category, where the entity text begins and ends, and the level of confidence that Amazon Comprehend Medical has in its detection.

detectPHIResponse_modelVersion :: Lens' DetectPHIResponse Text

The version of the model used to analyze the documents. The version number looks like X.X.X. You can use this information to track the model used for a particular batch of documents.