Network.AWS.ComprehendMedical.Lens

Description

Operations

DescribeEntitiesDetectionV2Job

describeEntitiesDetectionV2Job_jobId :: Lens' DescribeEntitiesDetectionV2Job Text

The identifier that Amazon Comprehend Medical generated for the job. The StartEntitiesDetectionV2Job operation returns this identifier in its response.

describeEntitiesDetectionV2JobResponse_comprehendMedicalAsyncJobProperties :: Lens' DescribeEntitiesDetectionV2JobResponse (Maybe ComprehendMedicalAsyncJobProperties)

An object that contains the properties associated with a detection job.

describeEntitiesDetectionV2JobResponse_httpStatus :: Lens' DescribeEntitiesDetectionV2JobResponse Int

The response's http status code.

DescribePHIDetectionJob

describePHIDetectionJob_jobId :: Lens' DescribePHIDetectionJob Text

The identifier that Amazon Comprehend Medical generated for the job. The StartPHIDetectionJob operation returns this identifier in its response.

describePHIDetectionJobResponse_comprehendMedicalAsyncJobProperties :: Lens' DescribePHIDetectionJobResponse (Maybe ComprehendMedicalAsyncJobProperties)

An object that contains the properties associated with a detection job.

describePHIDetectionJobResponse_httpStatus :: Lens' DescribePHIDetectionJobResponse Int

The response's http status code.

ListICD10CMInferenceJobs

listICD10CMInferenceJobs_nextToken :: Lens' ListICD10CMInferenceJobs (Maybe Text)

Identifies the next page of results to return.

listICD10CMInferenceJobs_filter :: Lens' ListICD10CMInferenceJobs (Maybe ComprehendMedicalAsyncJobFilter)

Filters the jobs that are returned. You can filter jobs based on their names, status, or the date and time that they were submitted. You can only set one filter at a time.

listICD10CMInferenceJobs_maxResults :: Lens' ListICD10CMInferenceJobs (Maybe Natural)

The maximum number of results to return in each page. The default is 100.

listICD10CMInferenceJobsResponse_nextToken :: Lens' ListICD10CMInferenceJobsResponse (Maybe Text)

Identifies the next page of results to return.

listICD10CMInferenceJobsResponse_comprehendMedicalAsyncJobPropertiesList :: Lens' ListICD10CMInferenceJobsResponse (Maybe [ComprehendMedicalAsyncJobProperties])

A list containing the properties of each job that is returned.

listICD10CMInferenceJobsResponse_httpStatus :: Lens' ListICD10CMInferenceJobsResponse Int

The response's http status code.

InferICD10CM

inferICD10CM_text :: Lens' InferICD10CM Text

The input text used for analysis. The input for InferICD10CM is a string from 1 to 10000 characters.

inferICD10CMResponse_paginationToken :: Lens' InferICD10CMResponse (Maybe Text)

If the result of the previous request to InferICD10CM was truncated, include the PaginationToken to fetch the next page of medical condition entities.

inferICD10CMResponse_modelVersion :: Lens' InferICD10CMResponse (Maybe Text)

The version of the model used to analyze the documents, in the format n.n.n You can use this information to track the model used for a particular batch of documents.

inferICD10CMResponse_httpStatus :: Lens' InferICD10CMResponse Int

The response's http status code.

inferICD10CMResponse_entities :: Lens' InferICD10CMResponse [ICD10CMEntity]

The medical conditions detected in the text linked to ICD-10-CM concepts. If the action is successful, the service sends back an HTTP 200 response, as well as the entities detected.

InferRxNorm

inferRxNorm_text :: Lens' InferRxNorm Text

The input text used for analysis. The input for InferRxNorm is a string from 1 to 10000 characters.

inferRxNormResponse_paginationToken :: Lens' InferRxNormResponse (Maybe Text)

If the result of the previous request to InferRxNorm was truncated, include the PaginationToken to fetch the next page of medication entities.

inferRxNormResponse_modelVersion :: Lens' InferRxNormResponse (Maybe Text)

The version of the model used to analyze the documents, in the format n.n.n You can use this information to track the model used for a particular batch of documents.

inferRxNormResponse_httpStatus :: Lens' InferRxNormResponse Int

The response's http status code.

inferRxNormResponse_entities :: Lens' InferRxNormResponse [RxNormEntity]

The medication entities detected in the text linked to RxNorm concepts. If the action is successful, the service sends back an HTTP 200 response, as well as the entities detected.

StartICD10CMInferenceJob

startICD10CMInferenceJob_kmsKey :: Lens' StartICD10CMInferenceJob (Maybe Text)

An AWS Key Management Service key to encrypt your output files. If you do not specify a key, the files are written in plain text.

startICD10CMInferenceJob_jobName :: Lens' StartICD10CMInferenceJob (Maybe Text)

The identifier of the job.

startICD10CMInferenceJob_clientRequestToken :: Lens' StartICD10CMInferenceJob (Maybe Text)

A unique identifier for the request. If you don't set the client request token, Amazon Comprehend Medical generates one.

startICD10CMInferenceJob_inputDataConfig :: Lens' StartICD10CMInferenceJob InputDataConfig

Specifies the format and location of the input data for the job.

startICD10CMInferenceJob_outputDataConfig :: Lens' StartICD10CMInferenceJob OutputDataConfig

Specifies where to send the output files.

startICD10CMInferenceJob_dataAccessRoleArn :: Lens' StartICD10CMInferenceJob Text

The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role that grants Amazon Comprehend Medical read access to your input data. For more information, see Role-Based Permissions Required for Asynchronous Operations.

startICD10CMInferenceJob_languageCode :: Lens' StartICD10CMInferenceJob LanguageCode

The language of the input documents. All documents must be in the same language.

startICD10CMInferenceJobResponse_jobId :: Lens' StartICD10CMInferenceJobResponse (Maybe Text)

The identifier generated for the job. To get the status of a job, use this identifier with the StartICD10CMInferenceJob operation.

startICD10CMInferenceJobResponse_httpStatus :: Lens' StartICD10CMInferenceJobResponse Int

The response's http status code.

StartRxNormInferenceJob

startRxNormInferenceJob_kmsKey :: Lens' StartRxNormInferenceJob (Maybe Text)

An AWS Key Management Service key to encrypt your output files. If you do not specify a key, the files are written in plain text.

startRxNormInferenceJob_jobName :: Lens' StartRxNormInferenceJob (Maybe Text)

The identifier of the job.

startRxNormInferenceJob_clientRequestToken :: Lens' StartRxNormInferenceJob (Maybe Text)

A unique identifier for the request. If you don't set the client request token, Amazon Comprehend Medical generates one.

startRxNormInferenceJob_inputDataConfig :: Lens' StartRxNormInferenceJob InputDataConfig

Specifies the format and location of the input data for the job.

startRxNormInferenceJob_outputDataConfig :: Lens' StartRxNormInferenceJob OutputDataConfig

Specifies where to send the output files.

startRxNormInferenceJob_dataAccessRoleArn :: Lens' StartRxNormInferenceJob Text

The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role that grants Amazon Comprehend Medical read access to your input data. For more information, see Role-Based Permissions Required for Asynchronous Operations.

startRxNormInferenceJob_languageCode :: Lens' StartRxNormInferenceJob LanguageCode

The language of the input documents. All documents must be in the same language.

startRxNormInferenceJobResponse_jobId :: Lens' StartRxNormInferenceJobResponse (Maybe Text)

The identifier of the job.

startRxNormInferenceJobResponse_httpStatus :: Lens' StartRxNormInferenceJobResponse Int

The response's http status code.

ListPHIDetectionJobs

listPHIDetectionJobs_nextToken :: Lens' ListPHIDetectionJobs (Maybe Text)

Identifies the next page of results to return.

listPHIDetectionJobs_filter :: Lens' ListPHIDetectionJobs (Maybe ComprehendMedicalAsyncJobFilter)

Filters the jobs that are returned. You can filter jobs based on their names, status, or the date and time that they were submitted. You can only set one filter at a time.

listPHIDetectionJobs_maxResults :: Lens' ListPHIDetectionJobs (Maybe Natural)

The maximum number of results to return in each page. The default is 100.

listPHIDetectionJobsResponse_nextToken :: Lens' ListPHIDetectionJobsResponse (Maybe Text)

Identifies the next page of results to return.

listPHIDetectionJobsResponse_comprehendMedicalAsyncJobPropertiesList :: Lens' ListPHIDetectionJobsResponse (Maybe [ComprehendMedicalAsyncJobProperties])

A list containing the properties of each job returned.

listPHIDetectionJobsResponse_httpStatus :: Lens' ListPHIDetectionJobsResponse Int

The response's http status code.

DescribeICD10CMInferenceJob

describeICD10CMInferenceJob_jobId :: Lens' DescribeICD10CMInferenceJob Text

The identifier that Amazon Comprehend Medical generated for the job. The StartICD10CMInferenceJob operation returns this identifier in its response.

describeICD10CMInferenceJobResponse_comprehendMedicalAsyncJobProperties :: Lens' DescribeICD10CMInferenceJobResponse (Maybe ComprehendMedicalAsyncJobProperties)

An object that contains the properties associated with a detection job.

describeICD10CMInferenceJobResponse_httpStatus :: Lens' DescribeICD10CMInferenceJobResponse Int

The response's http status code.

StartPHIDetectionJob

startPHIDetectionJob_kmsKey :: Lens' StartPHIDetectionJob (Maybe Text)

An AWS Key Management Service key to encrypt your output files. If you do not specify a key, the files are written in plain text.

startPHIDetectionJob_jobName :: Lens' StartPHIDetectionJob (Maybe Text)

The identifier of the job.

startPHIDetectionJob_clientRequestToken :: Lens' StartPHIDetectionJob (Maybe Text)

A unique identifier for the request. If you don't set the client request token, Amazon Comprehend Medical generates one.

startPHIDetectionJob_inputDataConfig :: Lens' StartPHIDetectionJob InputDataConfig

Specifies the format and location of the input data for the job.

startPHIDetectionJob_outputDataConfig :: Lens' StartPHIDetectionJob OutputDataConfig

Specifies where to send the output files.

startPHIDetectionJob_dataAccessRoleArn :: Lens' StartPHIDetectionJob Text

The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role that grants Amazon Comprehend Medical read access to your input data. For more information, see Role-Based Permissions Required for Asynchronous Operations.

startPHIDetectionJob_languageCode :: Lens' StartPHIDetectionJob LanguageCode

The language of the input documents. All documents must be in the same language.

startPHIDetectionJobResponse_jobId :: Lens' StartPHIDetectionJobResponse (Maybe Text)

The identifier generated for the job. To get the status of a job, use this identifier with the DescribePHIDetectionJob operation.

startPHIDetectionJobResponse_httpStatus :: Lens' StartPHIDetectionJobResponse Int

The response's http status code.

StopEntitiesDetectionV2Job

stopEntitiesDetectionV2Job_jobId :: Lens' StopEntitiesDetectionV2Job Text

The identifier of the medical entities job to stop.

stopEntitiesDetectionV2JobResponse_jobId :: Lens' StopEntitiesDetectionV2JobResponse (Maybe Text)

The identifier of the medical entities detection job that was stopped.

stopEntitiesDetectionV2JobResponse_httpStatus :: Lens' StopEntitiesDetectionV2JobResponse Int

The response's http status code.

DescribeRxNormInferenceJob

describeRxNormInferenceJob_jobId :: Lens' DescribeRxNormInferenceJob Text

The identifier that Amazon Comprehend Medical generated for the job. The StartRxNormInferenceJob operation returns this identifier in its response.

describeRxNormInferenceJobResponse_comprehendMedicalAsyncJobProperties :: Lens' DescribeRxNormInferenceJobResponse (Maybe ComprehendMedicalAsyncJobProperties)

An object that contains the properties associated with a detection job.

describeRxNormInferenceJobResponse_httpStatus :: Lens' DescribeRxNormInferenceJobResponse Int

The response's http status code.

StopICD10CMInferenceJob

stopICD10CMInferenceJob_jobId :: Lens' StopICD10CMInferenceJob Text

The identifier of the job.

stopICD10CMInferenceJobResponse_jobId :: Lens' StopICD10CMInferenceJobResponse (Maybe Text)

The identifier generated for the job. To get the status of job, use this identifier with the DescribeICD10CMInferenceJob operation.

stopICD10CMInferenceJobResponse_httpStatus :: Lens' StopICD10CMInferenceJobResponse Int

The response's http status code.

ListEntitiesDetectionV2Jobs

listEntitiesDetectionV2Jobs_nextToken :: Lens' ListEntitiesDetectionV2Jobs (Maybe Text)

Identifies the next page of results to return.

listEntitiesDetectionV2Jobs_filter :: Lens' ListEntitiesDetectionV2Jobs (Maybe ComprehendMedicalAsyncJobFilter)

Filters the jobs that are returned. You can filter jobs based on their names, status, or the date and time that they were submitted. You can only set one filter at a time.

listEntitiesDetectionV2Jobs_maxResults :: Lens' ListEntitiesDetectionV2Jobs (Maybe Natural)

The maximum number of results to return in each page. The default is 100.

listEntitiesDetectionV2JobsResponse_nextToken :: Lens' ListEntitiesDetectionV2JobsResponse (Maybe Text)

Identifies the next page of results to return.

listEntitiesDetectionV2JobsResponse_comprehendMedicalAsyncJobPropertiesList :: Lens' ListEntitiesDetectionV2JobsResponse (Maybe [ComprehendMedicalAsyncJobProperties])

A list containing the properties of each job returned.

listEntitiesDetectionV2JobsResponse_httpStatus :: Lens' ListEntitiesDetectionV2JobsResponse Int

The response's http status code.

StopRxNormInferenceJob

stopRxNormInferenceJob_jobId :: Lens' StopRxNormInferenceJob Text

The identifier of the job.

stopRxNormInferenceJobResponse_jobId :: Lens' StopRxNormInferenceJobResponse (Maybe Text)

The identifier generated for the job. To get the status of job, use this identifier with the DescribeRxNormInferenceJob operation.

stopRxNormInferenceJobResponse_httpStatus :: Lens' StopRxNormInferenceJobResponse Int

The response's http status code.

DetectPHI

detectPHI_text :: Lens' DetectPHI Text

A UTF-8 text string containing the clinical content being examined for PHI entities. Each string must contain fewer than 20,000 bytes of characters.

detectPHIResponse_paginationToken :: Lens' DetectPHIResponse (Maybe Text)

If the result of the previous request to DetectPHI was truncated, include the PaginationToken to fetch the next page of PHI entities.

detectPHIResponse_httpStatus :: Lens' DetectPHIResponse Int

The response's http status code.

detectPHIResponse_entities :: Lens' DetectPHIResponse [Entity]

The collection of PHI entities extracted from the input text and their associated information. For each entity, the response provides the entity text, the entity category, where the entity text begins and ends, and the level of confidence that Amazon Comprehend Medical has in its detection.

detectPHIResponse_modelVersion :: Lens' DetectPHIResponse Text

The version of the model used to analyze the documents. The version number looks like X.X.X. You can use this information to track the model used for a particular batch of documents.

DetectEntitiesV2

detectEntitiesV2_text :: Lens' DetectEntitiesV2 Text

A UTF-8 string containing the clinical content being examined for entities. Each string must contain fewer than 20,000 bytes of characters.

detectEntitiesV2Response_paginationToken :: Lens' DetectEntitiesV2Response (Maybe Text)

If the result to the DetectEntitiesV2 operation was truncated, include the PaginationToken to fetch the next page of entities.

detectEntitiesV2Response_unmappedAttributes :: Lens' DetectEntitiesV2Response (Maybe [UnmappedAttribute])

Attributes extracted from the input text that couldn't be related to an entity.

detectEntitiesV2Response_httpStatus :: Lens' DetectEntitiesV2Response Int

The response's http status code.

detectEntitiesV2Response_entities :: Lens' DetectEntitiesV2Response [Entity]

The collection of medical entities extracted from the input text and their associated information. For each entity, the response provides the entity text, the entity category, where the entity text begins and ends, and the level of confidence in the detection and analysis. Attributes and traits of the entity are also returned.

detectEntitiesV2Response_modelVersion :: Lens' DetectEntitiesV2Response Text

The version of the model used to analyze the documents. The version number looks like X.X.X. You can use this information to track the model used for a particular batch of documents.

StopPHIDetectionJob

stopPHIDetectionJob_jobId :: Lens' StopPHIDetectionJob Text

The identifier of the PHI detection job to stop.

stopPHIDetectionJobResponse_jobId :: Lens' StopPHIDetectionJobResponse (Maybe Text)

The identifier of the PHI detection job that was stopped.

stopPHIDetectionJobResponse_httpStatus :: Lens' StopPHIDetectionJobResponse Int

The response's http status code.

StartEntitiesDetectionV2Job

startEntitiesDetectionV2Job_kmsKey :: Lens' StartEntitiesDetectionV2Job (Maybe Text)

An AWS Key Management Service key to encrypt your output files. If you do not specify a key, the files are written in plain text.

startEntitiesDetectionV2Job_jobName :: Lens' StartEntitiesDetectionV2Job (Maybe Text)

The identifier of the job.

startEntitiesDetectionV2Job_clientRequestToken :: Lens' StartEntitiesDetectionV2Job (Maybe Text)

A unique identifier for the request. If you don't set the client request token, Amazon Comprehend Medical generates one.

startEntitiesDetectionV2Job_inputDataConfig :: Lens' StartEntitiesDetectionV2Job InputDataConfig

Specifies the format and location of the input data for the job.

startEntitiesDetectionV2Job_outputDataConfig :: Lens' StartEntitiesDetectionV2Job OutputDataConfig

Specifies where to send the output files.

startEntitiesDetectionV2Job_dataAccessRoleArn :: Lens' StartEntitiesDetectionV2Job Text

The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role that grants Amazon Comprehend Medical read access to your input data. For more information, see Role-Based Permissions Required for Asynchronous Operations.

startEntitiesDetectionV2Job_languageCode :: Lens' StartEntitiesDetectionV2Job LanguageCode

The language of the input documents. All documents must be in the same language.

startEntitiesDetectionV2JobResponse_jobId :: Lens' StartEntitiesDetectionV2JobResponse (Maybe Text)

The identifier generated for the job. To get the status of a job, use this identifier with the DescribeEntitiesDetectionV2Job operation.

startEntitiesDetectionV2JobResponse_httpStatus :: Lens' StartEntitiesDetectionV2JobResponse Int

The response's http status code.

ListRxNormInferenceJobs

listRxNormInferenceJobs_nextToken :: Lens' ListRxNormInferenceJobs (Maybe Text)

Identifies the next page of results to return.

listRxNormInferenceJobs_filter :: Lens' ListRxNormInferenceJobs (Maybe ComprehendMedicalAsyncJobFilter)

Filters the jobs that are returned. You can filter jobs based on their names, status, or the date and time that they were submitted. You can only set one filter at a time.

listRxNormInferenceJobs_maxResults :: Lens' ListRxNormInferenceJobs (Maybe Natural)

Identifies the next page of results to return.

listRxNormInferenceJobsResponse_nextToken :: Lens' ListRxNormInferenceJobsResponse (Maybe Text)

Identifies the next page of results to return.

listRxNormInferenceJobsResponse_comprehendMedicalAsyncJobPropertiesList :: Lens' ListRxNormInferenceJobsResponse (Maybe [ComprehendMedicalAsyncJobProperties])

The maximum number of results to return in each page. The default is 100.

listRxNormInferenceJobsResponse_httpStatus :: Lens' ListRxNormInferenceJobsResponse Int

The response's http status code.

Types

Attribute

attribute_relationshipScore :: Lens' Attribute (Maybe Double)

The level of confidence that Amazon Comprehend Medical has that this attribute is correctly related to this entity.

attribute_beginOffset :: Lens' Attribute (Maybe Int)

The 0-based character offset in the input text that shows where the attribute begins. The offset returns the UTF-8 code point in the string.

attribute_text :: Lens' Attribute (Maybe Text)

The segment of input text extracted as this attribute.

attribute_category :: Lens' Attribute (Maybe EntityType)

The category of attribute.

attribute_score :: Lens' Attribute (Maybe Double)

The level of confidence that Amazon Comprehend Medical has that the segment of text is correctly recognized as an attribute.

attribute_traits :: Lens' Attribute (Maybe [Trait])

Contextual information for this attribute.

attribute_relationshipType :: Lens' Attribute (Maybe RelationshipType)

The type of relationship between the entity and attribute. Type for the relationship is OVERLAP, indicating that the entity occurred at the same time as the Date_Expression.

attribute_endOffset :: Lens' Attribute (Maybe Int)

The 0-based character offset in the input text that shows where the attribute ends. The offset returns the UTF-8 code point in the string.

attribute_id :: Lens' Attribute (Maybe Int)

The numeric identifier for this attribute. This is a monotonically increasing id unique within this response rather than a global unique identifier.

attribute_type :: Lens' Attribute (Maybe EntitySubType)

The type of attribute.

ComprehendMedicalAsyncJobFilter

comprehendMedicalAsyncJobFilter_submitTimeAfter :: Lens' ComprehendMedicalAsyncJobFilter (Maybe UTCTime)

Filters the list of jobs based on the time that the job was submitted for processing. Returns only jobs submitted after the specified time. Jobs are returned in descending order, newest to oldest.

comprehendMedicalAsyncJobFilter_submitTimeBefore :: Lens' ComprehendMedicalAsyncJobFilter (Maybe UTCTime)

Filters the list of jobs based on the time that the job was submitted for processing. Returns only jobs submitted before the specified time. Jobs are returned in ascending order, oldest to newest.

comprehendMedicalAsyncJobFilter_jobName :: Lens' ComprehendMedicalAsyncJobFilter (Maybe Text)

Filters on the name of the job.

comprehendMedicalAsyncJobFilter_jobStatus :: Lens' ComprehendMedicalAsyncJobFilter (Maybe JobStatus)

Filters the list of jobs based on job status. Returns only jobs with the specified status.

ComprehendMedicalAsyncJobProperties

comprehendMedicalAsyncJobProperties_languageCode :: Lens' ComprehendMedicalAsyncJobProperties (Maybe LanguageCode)

The language code of the input documents.

comprehendMedicalAsyncJobProperties_kmsKey :: Lens' ComprehendMedicalAsyncJobProperties (Maybe Text)

The AWS Key Management Service key, if any, used to encrypt the output files.

comprehendMedicalAsyncJobProperties_jobId :: Lens' ComprehendMedicalAsyncJobProperties (Maybe Text)

The identifier assigned to the detection job.

comprehendMedicalAsyncJobProperties_jobName :: Lens' ComprehendMedicalAsyncJobProperties (Maybe Text)

The name that you assigned to the detection job.

comprehendMedicalAsyncJobProperties_manifestFilePath :: Lens' ComprehendMedicalAsyncJobProperties (Maybe Text)

The path to the file that describes the results of a batch job.

comprehendMedicalAsyncJobProperties_inputDataConfig :: Lens' ComprehendMedicalAsyncJobProperties (Maybe InputDataConfig)

The input data configuration that you supplied when you created the detection job.

comprehendMedicalAsyncJobProperties_endTime :: Lens' ComprehendMedicalAsyncJobProperties (Maybe UTCTime)

The time that the detection job completed.

comprehendMedicalAsyncJobProperties_outputDataConfig :: Lens' ComprehendMedicalAsyncJobProperties (Maybe OutputDataConfig)

The output data configuration that you supplied when you created the detection job.

comprehendMedicalAsyncJobProperties_modelVersion :: Lens' ComprehendMedicalAsyncJobProperties (Maybe Text)

The version of the model used to analyze the documents. The version number looks like X.X.X. You can use this information to track the model used for a particular batch of documents.

comprehendMedicalAsyncJobProperties_dataAccessRoleArn :: Lens' ComprehendMedicalAsyncJobProperties (Maybe Text)

The Amazon Resource Name (ARN) that gives Amazon Comprehend Medical read access to your input data.

comprehendMedicalAsyncJobProperties_jobStatus :: Lens' ComprehendMedicalAsyncJobProperties (Maybe JobStatus)

The current status of the detection job. If the status is FAILED, the Message field shows the reason for the failure.

comprehendMedicalAsyncJobProperties_message :: Lens' ComprehendMedicalAsyncJobProperties (Maybe Text)

A description of the status of a job.

comprehendMedicalAsyncJobProperties_expirationTime :: Lens' ComprehendMedicalAsyncJobProperties (Maybe UTCTime)

The date and time that job metadata is deleted from the server. Output files in your S3 bucket will not be deleted. After the metadata is deleted, the job will no longer appear in the results of the ListEntitiesDetectionV2Job or the ListPHIDetectionJobs operation.

comprehendMedicalAsyncJobProperties_submitTime :: Lens' ComprehendMedicalAsyncJobProperties (Maybe UTCTime)

The time that the detection job was submitted for processing.

Entity

entity_beginOffset :: Lens' Entity (Maybe Int)

The 0-based character offset in the input text that shows where the entity begins. The offset returns the UTF-8 code point in the string.

entity_text :: Lens' Entity (Maybe Text)

The segment of input text extracted as this entity.

entity_category :: Lens' Entity (Maybe EntityType)

The category of the entity.

entity_score :: Lens' Entity (Maybe Double)

The level of confidence that Amazon Comprehend Medical has in the accuracy of the detection.

entity_traits :: Lens' Entity (Maybe [Trait])

Contextual information for the entity.

entity_attributes :: Lens' Entity (Maybe [Attribute])

The extracted attributes that relate to this entity.

entity_endOffset :: Lens' Entity (Maybe Int)

The 0-based character offset in the input text that shows where the entity ends. The offset returns the UTF-8 code point in the string.

entity_id :: Lens' Entity (Maybe Int)

The numeric identifier for the entity. This is a monotonically increasing id unique within this response rather than a global unique identifier.

entity_type :: Lens' Entity (Maybe EntitySubType)

Describes the specific type of entity with category of entities.

ICD10CMAttribute

iCD10CMAttribute_relationshipScore :: Lens' ICD10CMAttribute (Maybe Double)

The level of confidence that Amazon Comprehend Medical has that this attribute is correctly related to this entity.

iCD10CMAttribute_beginOffset :: Lens' ICD10CMAttribute (Maybe Int)

The 0-based character offset in the input text that shows where the attribute begins. The offset returns the UTF-8 code point in the string.

iCD10CMAttribute_text :: Lens' ICD10CMAttribute (Maybe Text)

The segment of input text which contains the detected attribute.

iCD10CMAttribute_category :: Lens' ICD10CMAttribute (Maybe ICD10CMEntityType)

The category of attribute. Can be either of DX_NAME or TIME_EXPRESSION.

iCD10CMAttribute_score :: Lens' ICD10CMAttribute (Maybe Double)

The level of confidence that Amazon Comprehend Medical has that the segment of text is correctly recognized as an attribute.

iCD10CMAttribute_traits :: Lens' ICD10CMAttribute (Maybe [ICD10CMTrait])

The contextual information for the attribute. The traits recognized by InferICD10CM are DIAGNOSIS, SIGN, SYMPTOM, and NEGATION.

iCD10CMAttribute_relationshipType :: Lens' ICD10CMAttribute (Maybe ICD10CMRelationshipType)

The type of relationship between the entity and attribute. Type for the relationship can be either of OVERLAP or SYSTEM_ORGAN_SITE.

iCD10CMAttribute_endOffset :: Lens' ICD10CMAttribute (Maybe Int)

The 0-based character offset in the input text that shows where the attribute ends. The offset returns the UTF-8 code point in the string.

iCD10CMAttribute_id :: Lens' ICD10CMAttribute (Maybe Int)

The numeric identifier for this attribute. This is a monotonically increasing id unique within this response rather than a global unique identifier.

iCD10CMAttribute_type :: Lens' ICD10CMAttribute (Maybe ICD10CMAttributeType)

The type of attribute. InferICD10CM detects entities of the type DX_NAME.

ICD10CMConcept

iCD10CMConcept_score :: Lens' ICD10CMConcept (Maybe Double)

The level of confidence that Amazon Comprehend Medical has that the entity is accurately linked to an ICD-10-CM concept.

iCD10CMConcept_code :: Lens' ICD10CMConcept (Maybe Text)

The ICD-10-CM code that identifies the concept found in the knowledge base from the Centers for Disease Control.

iCD10CMConcept_description :: Lens' ICD10CMConcept (Maybe Text)

The long description of the ICD-10-CM code in the ontology.

ICD10CMEntity

iCD10CMEntity_iCD10CMConcepts :: Lens' ICD10CMEntity (Maybe [ICD10CMConcept])

The ICD-10-CM concepts that the entity could refer to, along with a score indicating the likelihood of the match.

iCD10CMEntity_beginOffset :: Lens' ICD10CMEntity (Maybe Int)

The 0-based character offset in the input text that shows where the entity begins. The offset returns the UTF-8 code point in the string.

iCD10CMEntity_text :: Lens' ICD10CMEntity (Maybe Text)

The segment of input text that is matched to the detected entity.

iCD10CMEntity_category :: Lens' ICD10CMEntity (Maybe ICD10CMEntityCategory)

The category of the entity. InferICD10CM detects entities in the MEDICAL_CONDITION category.

iCD10CMEntity_score :: Lens' ICD10CMEntity (Maybe Double)

The level of confidence that Amazon Comprehend Medical has in the accuracy of the detection.

iCD10CMEntity_traits :: Lens' ICD10CMEntity (Maybe [ICD10CMTrait])

Provides Contextual information for the entity. The traits recognized by InferICD10CM are DIAGNOSIS, SIGN, SYMPTOM, and NEGATION.

iCD10CMEntity_attributes :: Lens' ICD10CMEntity (Maybe [ICD10CMAttribute])

The detected attributes that relate to the entity. An extracted segment of the text that is an attribute of an entity, or otherwise related to an entity, such as the nature of a medical condition.

iCD10CMEntity_endOffset :: Lens' ICD10CMEntity (Maybe Int)

The 0-based character offset in the input text that shows where the entity ends. The offset returns the UTF-8 code point in the string.

iCD10CMEntity_id :: Lens' ICD10CMEntity (Maybe Int)

The numeric identifier for the entity. This is a monotonically increasing id unique within this response rather than a global unique identifier.

iCD10CMEntity_type :: Lens' ICD10CMEntity (Maybe ICD10CMEntityType)

Describes the specific type of entity with category of entities. InferICD10CM detects entities of the type DX_NAME and TIME_EXPRESSION.

ICD10CMTrait

iCD10CMTrait_score :: Lens' ICD10CMTrait (Maybe Double)

The level of confidence that Amazon Comprehend Medical has that the segment of text is correctly recognized as a trait.

iCD10CMTrait_name :: Lens' ICD10CMTrait (Maybe ICD10CMTraitName)

Provides a name or contextual description about the trait.

InputDataConfig

inputDataConfig_s3Key :: Lens' InputDataConfig (Maybe Text)

The path to the input data files in the S3 bucket.

inputDataConfig_s3Bucket :: Lens' InputDataConfig Text

The URI of the S3 bucket that contains the input data. The bucket must be in the same region as the API endpoint that you are calling.

Each file in the document collection must be less than 40 KB. You can store a maximum of 30 GB in the bucket.

OutputDataConfig

outputDataConfig_s3Key :: Lens' OutputDataConfig (Maybe Text)

The path to the output data files in the S3 bucket. Amazon Comprehend Medical creates an output directory using the job ID so that the output from one job does not overwrite the output of another.

outputDataConfig_s3Bucket :: Lens' OutputDataConfig Text

When you use the OutputDataConfig object with asynchronous operations, you specify the Amazon S3 location where you want to write the output data. The URI must be in the same region as the API endpoint that you are calling. The location is used as the prefix for the actual location of the output.

RxNormAttribute

rxNormAttribute_relationshipScore :: Lens' RxNormAttribute (Maybe Double)

The level of confidence that Amazon Comprehend Medical has that the attribute is accurately linked to an entity.

rxNormAttribute_beginOffset :: Lens' RxNormAttribute (Maybe Int)

The 0-based character offset in the input text that shows where the attribute begins. The offset returns the UTF-8 code point in the string.

rxNormAttribute_text :: Lens' RxNormAttribute (Maybe Text)

The segment of input text which corresponds to the detected attribute.

rxNormAttribute_score :: Lens' RxNormAttribute (Maybe Double)

The level of confidence that Comprehend Medical has that the segment of text is correctly recognized as an attribute.

rxNormAttribute_traits :: Lens' RxNormAttribute (Maybe [RxNormTrait])

Contextual information for the attribute. InferRxNorm recognizes the trait NEGATION for attributes, i.e. that the patient is not taking a specific dose or form of a medication.

rxNormAttribute_endOffset :: Lens' RxNormAttribute (Maybe Int)

The 0-based character offset in the input text that shows where the attribute ends. The offset returns the UTF-8 code point in the string.

rxNormAttribute_id :: Lens' RxNormAttribute (Maybe Int)

The numeric identifier for this attribute. This is a monotonically increasing id unique within this response rather than a global unique identifier.

rxNormAttribute_type :: Lens' RxNormAttribute (Maybe RxNormAttributeType)

The type of attribute. The types of attributes recognized by InferRxNorm are BRAND_NAME and GENERIC_NAME.

RxNormConcept

rxNormConcept_score :: Lens' RxNormConcept (Maybe Double)

The level of confidence that Amazon Comprehend Medical has that the entity is accurately linked to the reported RxNorm concept.

rxNormConcept_code :: Lens' RxNormConcept (Maybe Text)

RxNorm concept ID, also known as the RxCUI.

rxNormConcept_description :: Lens' RxNormConcept (Maybe Text)

The description of the RxNorm concept.

RxNormEntity

rxNormEntity_rxNormConcepts :: Lens' RxNormEntity (Maybe [RxNormConcept])

The RxNorm concepts that the entity could refer to, along with a score indicating the likelihood of the match.

rxNormEntity_beginOffset :: Lens' RxNormEntity (Maybe Int)

The 0-based character offset in the input text that shows where the entity begins. The offset returns the UTF-8 code point in the string.

rxNormEntity_text :: Lens' RxNormEntity (Maybe Text)

The segment of input text extracted from which the entity was detected.

rxNormEntity_category :: Lens' RxNormEntity (Maybe RxNormEntityCategory)

The category of the entity. The recognized categories are GENERIC or BRAND_NAME.

rxNormEntity_score :: Lens' RxNormEntity (Maybe Double)

The level of confidence that Amazon Comprehend Medical has in the accuracy of the detected entity.

rxNormEntity_traits :: Lens' RxNormEntity (Maybe [RxNormTrait])

Contextual information for the entity.

rxNormEntity_attributes :: Lens' RxNormEntity (Maybe [RxNormAttribute])

The extracted attributes that relate to the entity. The attributes recognized by InferRxNorm are DOSAGE, DURATION, FORM, FREQUENCY, RATE, ROUTE_OR_MODE, and STRENGTH.

rxNormEntity_endOffset :: Lens' RxNormEntity (Maybe Int)

The 0-based character offset in the input text that shows where the entity ends. The offset returns the UTF-8 code point in the string.

rxNormEntity_id :: Lens' RxNormEntity (Maybe Int)

The numeric identifier for the entity. This is a monotonically increasing id unique within this response rather than a global unique identifier.

rxNormEntity_type :: Lens' RxNormEntity (Maybe RxNormEntityType)

Describes the specific type of entity. For InferRxNorm, the recognized entity type is MEDICATION.

RxNormTrait

rxNormTrait_score :: Lens' RxNormTrait (Maybe Double)

The level of confidence that Amazon Comprehend Medical has in the accuracy of the detected trait.

rxNormTrait_name :: Lens' RxNormTrait (Maybe RxNormTraitName)

Provides a name or contextual description about the trait.

Trait

trait_score :: Lens' Trait (Maybe Double)

The level of confidence that Amazon Comprehend Medical has in the accuracy of this trait.

trait_name :: Lens' Trait (Maybe AttributeName)

Provides a name or contextual description about the trait.

UnmappedAttribute

unmappedAttribute_attribute :: Lens' UnmappedAttribute (Maybe Attribute)

The specific attribute that has been extracted but not mapped to an entity.

unmappedAttribute_type :: Lens' UnmappedAttribute (Maybe EntityType)

The type of the attribute, could be one of the following values: "MEDICATION", "MEDICAL_CONDITION", "ANATOMY", "TEST_AND_TREATMENT_PROCEDURE" or "PROTECTED_HEALTH_INFORMATION".