Network.AWS.ComprehendMedical.ListICD10CMInferenceJobs

Description

Gets a list of InferICD10CM jobs that you have submitted.

Creating a Request

data ListICD10CMInferenceJobs

See: newListICD10CMInferenceJobs smart constructor.

Constructors

ListICD10CMInferenceJobs'

Fields

nextToken :: Maybe Text

Identifies the next page of results to return.

filter' :: Maybe ComprehendMedicalAsyncJobFilter

Filters the jobs that are returned. You can filter jobs based on their names, status, or the date and time that they were submitted. You can only set one filter at a time.

maxResults :: Maybe Natural

The maximum number of results to return in each page. The default is 100.

newListICD10CMInferenceJobs :: ListICD10CMInferenceJobs

Create a value of ListICD10CMInferenceJobs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListICD10CMInferenceJobs', listICD10CMInferenceJobs_nextToken - Identifies the next page of results to return.

$sel:filter':ListICD10CMInferenceJobs', listICD10CMInferenceJobs_filter - Filters the jobs that are returned. You can filter jobs based on their names, status, or the date and time that they were submitted. You can only set one filter at a time.

$sel:maxResults:ListICD10CMInferenceJobs', listICD10CMInferenceJobs_maxResults - The maximum number of results to return in each page. The default is 100.

Request Lenses

listICD10CMInferenceJobs_nextToken :: Lens' ListICD10CMInferenceJobs (Maybe Text)

Identifies the next page of results to return.

listICD10CMInferenceJobs_filter :: Lens' ListICD10CMInferenceJobs (Maybe ComprehendMedicalAsyncJobFilter)

Filters the jobs that are returned. You can filter jobs based on their names, status, or the date and time that they were submitted. You can only set one filter at a time.

listICD10CMInferenceJobs_maxResults :: Lens' ListICD10CMInferenceJobs (Maybe Natural)

The maximum number of results to return in each page. The default is 100.

Destructuring the Response

data ListICD10CMInferenceJobsResponse

See: newListICD10CMInferenceJobsResponse smart constructor.

Constructors

ListICD10CMInferenceJobsResponse'

Fields

nextToken :: Maybe Text

Identifies the next page of results to return.

comprehendMedicalAsyncJobPropertiesList :: Maybe [ComprehendMedicalAsyncJobProperties]

A list containing the properties of each job that is returned.

httpStatus :: Int

The response's http status code.

newListICD10CMInferenceJobsResponse

Arguments

:: Int
$sel:httpStatus:ListICD10CMInferenceJobsResponse'

-> ListICD10CMInferenceJobsResponse

Create a value of ListICD10CMInferenceJobsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListICD10CMInferenceJobs', listICD10CMInferenceJobsResponse_nextToken - Identifies the next page of results to return.

$sel:comprehendMedicalAsyncJobPropertiesList:ListICD10CMInferenceJobsResponse', listICD10CMInferenceJobsResponse_comprehendMedicalAsyncJobPropertiesList - A list containing the properties of each job that is returned.

$sel:httpStatus:ListICD10CMInferenceJobsResponse', listICD10CMInferenceJobsResponse_httpStatus - The response's http status code.

Response Lenses

listICD10CMInferenceJobsResponse_nextToken :: Lens' ListICD10CMInferenceJobsResponse (Maybe Text)

Identifies the next page of results to return.

listICD10CMInferenceJobsResponse_comprehendMedicalAsyncJobPropertiesList :: Lens' ListICD10CMInferenceJobsResponse (Maybe [ComprehendMedicalAsyncJobProperties])

A list containing the properties of each job that is returned.

listICD10CMInferenceJobsResponse_httpStatus :: Lens' ListICD10CMInferenceJobsResponse Int

The response's http status code.