Network.AWS.ComprehendMedical.ListRxNormInferenceJobs

Description

Gets a list of InferRxNorm jobs that you have submitted.

Creating a Request

data ListRxNormInferenceJobs

See: newListRxNormInferenceJobs smart constructor.

Constructors

ListRxNormInferenceJobs'

Fields

nextToken :: Maybe Text

Identifies the next page of results to return.

filter' :: Maybe ComprehendMedicalAsyncJobFilter

Filters the jobs that are returned. You can filter jobs based on their names, status, or the date and time that they were submitted. You can only set one filter at a time.

maxResults :: Maybe Natural

Identifies the next page of results to return.

newListRxNormInferenceJobs :: ListRxNormInferenceJobs

Create a value of ListRxNormInferenceJobs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListRxNormInferenceJobs', listRxNormInferenceJobs_nextToken - Identifies the next page of results to return.

$sel:filter':ListRxNormInferenceJobs', listRxNormInferenceJobs_filter - Filters the jobs that are returned. You can filter jobs based on their names, status, or the date and time that they were submitted. You can only set one filter at a time.

$sel:maxResults:ListRxNormInferenceJobs', listRxNormInferenceJobs_maxResults - Identifies the next page of results to return.

Request Lenses

listRxNormInferenceJobs_nextToken :: Lens' ListRxNormInferenceJobs (Maybe Text)

Identifies the next page of results to return.

listRxNormInferenceJobs_filter :: Lens' ListRxNormInferenceJobs (Maybe ComprehendMedicalAsyncJobFilter)

Filters the jobs that are returned. You can filter jobs based on their names, status, or the date and time that they were submitted. You can only set one filter at a time.

listRxNormInferenceJobs_maxResults :: Lens' ListRxNormInferenceJobs (Maybe Natural)

Identifies the next page of results to return.

Destructuring the Response

data ListRxNormInferenceJobsResponse

See: newListRxNormInferenceJobsResponse smart constructor.

Constructors

ListRxNormInferenceJobsResponse'

Fields

nextToken :: Maybe Text

Identifies the next page of results to return.

comprehendMedicalAsyncJobPropertiesList :: Maybe [ComprehendMedicalAsyncJobProperties]

The maximum number of results to return in each page. The default is 100.

httpStatus :: Int

The response's http status code.

newListRxNormInferenceJobsResponse

Arguments

:: Int
$sel:httpStatus:ListRxNormInferenceJobsResponse'

-> ListRxNormInferenceJobsResponse

Create a value of ListRxNormInferenceJobsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListRxNormInferenceJobs', listRxNormInferenceJobsResponse_nextToken - Identifies the next page of results to return.

$sel:comprehendMedicalAsyncJobPropertiesList:ListRxNormInferenceJobsResponse', listRxNormInferenceJobsResponse_comprehendMedicalAsyncJobPropertiesList - The maximum number of results to return in each page. The default is 100.

$sel:httpStatus:ListRxNormInferenceJobsResponse', listRxNormInferenceJobsResponse_httpStatus - The response's http status code.

Response Lenses

listRxNormInferenceJobsResponse_nextToken :: Lens' ListRxNormInferenceJobsResponse (Maybe Text)

Identifies the next page of results to return.

listRxNormInferenceJobsResponse_comprehendMedicalAsyncJobPropertiesList :: Lens' ListRxNormInferenceJobsResponse (Maybe [ComprehendMedicalAsyncJobProperties])

The maximum number of results to return in each page. The default is 100.

listRxNormInferenceJobsResponse_httpStatus :: Lens' ListRxNormInferenceJobsResponse Int

The response's http status code.