Network.AWS.ComprehendMedical.StartEntitiesDetectionV2Job

Description

Starts an asynchronous medical entity detection job for a collection of documents. Use the DescribeEntitiesDetectionV2Job operation to track the status of a job.

Creating a Request

data StartEntitiesDetectionV2Job

See: newStartEntitiesDetectionV2Job smart constructor.

Constructors

StartEntitiesDetectionV2Job'

Fields

kmsKey :: Maybe Text

An AWS Key Management Service key to encrypt your output files. If you do not specify a key, the files are written in plain text.

jobName :: Maybe Text

The identifier of the job.

clientRequestToken :: Maybe Text

A unique identifier for the request. If you don't set the client request token, Amazon Comprehend Medical generates one.

inputDataConfig :: InputDataConfig

Specifies the format and location of the input data for the job.

outputDataConfig :: OutputDataConfig

Specifies where to send the output files.

dataAccessRoleArn :: Text

The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role that grants Amazon Comprehend Medical read access to your input data. For more information, see Role-Based Permissions Required for Asynchronous Operations.

languageCode :: LanguageCode

The language of the input documents. All documents must be in the same language.

newStartEntitiesDetectionV2Job

Arguments

:: InputDataConfig
$sel:inputDataConfig:StartEntitiesDetectionV2Job'

-> OutputDataConfig
$sel:outputDataConfig:StartEntitiesDetectionV2Job'

-> Text
$sel:dataAccessRoleArn:StartEntitiesDetectionV2Job'

-> LanguageCode
$sel:languageCode:StartEntitiesDetectionV2Job'

-> StartEntitiesDetectionV2Job

Create a value of StartEntitiesDetectionV2Job with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:kmsKey:StartEntitiesDetectionV2Job', startEntitiesDetectionV2Job_kmsKey - An AWS Key Management Service key to encrypt your output files. If you do not specify a key, the files are written in plain text.

$sel:jobName:StartEntitiesDetectionV2Job', startEntitiesDetectionV2Job_jobName - The identifier of the job.

$sel:clientRequestToken:StartEntitiesDetectionV2Job', startEntitiesDetectionV2Job_clientRequestToken - A unique identifier for the request. If you don't set the client request token, Amazon Comprehend Medical generates one.

$sel:inputDataConfig:StartEntitiesDetectionV2Job', startEntitiesDetectionV2Job_inputDataConfig - Specifies the format and location of the input data for the job.

$sel:outputDataConfig:StartEntitiesDetectionV2Job', startEntitiesDetectionV2Job_outputDataConfig - Specifies where to send the output files.

$sel:dataAccessRoleArn:StartEntitiesDetectionV2Job', startEntitiesDetectionV2Job_dataAccessRoleArn - The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role that grants Amazon Comprehend Medical read access to your input data. For more information, see Role-Based Permissions Required for Asynchronous Operations.

$sel:languageCode:StartEntitiesDetectionV2Job', startEntitiesDetectionV2Job_languageCode - The language of the input documents. All documents must be in the same language.

Request Lenses

startEntitiesDetectionV2Job_kmsKey :: Lens' StartEntitiesDetectionV2Job (Maybe Text)

An AWS Key Management Service key to encrypt your output files. If you do not specify a key, the files are written in plain text.

startEntitiesDetectionV2Job_jobName :: Lens' StartEntitiesDetectionV2Job (Maybe Text)

The identifier of the job.

startEntitiesDetectionV2Job_clientRequestToken :: Lens' StartEntitiesDetectionV2Job (Maybe Text)

A unique identifier for the request. If you don't set the client request token, Amazon Comprehend Medical generates one.

startEntitiesDetectionV2Job_inputDataConfig :: Lens' StartEntitiesDetectionV2Job InputDataConfig

Specifies the format and location of the input data for the job.

startEntitiesDetectionV2Job_outputDataConfig :: Lens' StartEntitiesDetectionV2Job OutputDataConfig

Specifies where to send the output files.

startEntitiesDetectionV2Job_dataAccessRoleArn :: Lens' StartEntitiesDetectionV2Job Text

The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role that grants Amazon Comprehend Medical read access to your input data. For more information, see Role-Based Permissions Required for Asynchronous Operations.

startEntitiesDetectionV2Job_languageCode :: Lens' StartEntitiesDetectionV2Job LanguageCode

The language of the input documents. All documents must be in the same language.

Destructuring the Response

data StartEntitiesDetectionV2JobResponse

See: newStartEntitiesDetectionV2JobResponse smart constructor.

Constructors

StartEntitiesDetectionV2JobResponse'

Fields

jobId :: Maybe Text

The identifier generated for the job. To get the status of a job, use this identifier with the DescribeEntitiesDetectionV2Job operation.

httpStatus :: Int

The response's http status code.

newStartEntitiesDetectionV2JobResponse

Arguments

:: Int
$sel:httpStatus:StartEntitiesDetectionV2JobResponse'

-> StartEntitiesDetectionV2JobResponse

Create a value of StartEntitiesDetectionV2JobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobId:StartEntitiesDetectionV2JobResponse', startEntitiesDetectionV2JobResponse_jobId - The identifier generated for the job. To get the status of a job, use this identifier with the DescribeEntitiesDetectionV2Job operation.

$sel:httpStatus:StartEntitiesDetectionV2JobResponse', startEntitiesDetectionV2JobResponse_httpStatus - The response's http status code.

Response Lenses

startEntitiesDetectionV2JobResponse_jobId :: Lens' StartEntitiesDetectionV2JobResponse (Maybe Text)

The identifier generated for the job. To get the status of a job, use this identifier with the DescribeEntitiesDetectionV2Job operation.

startEntitiesDetectionV2JobResponse_httpStatus :: Lens' StartEntitiesDetectionV2JobResponse Int

The response's http status code.