Network.AWS.ComprehendMedical.Types.Attribute

Description

Documentation

data Attribute

An extracted segment of the text that is an attribute of an entity, or otherwise related to an entity, such as the dosage of a medication taken. It contains information about the attribute such as id, begin and end offset within the input text, and the segment of the input text.

See: newAttribute smart constructor.

Constructors

Attribute'

Fields

relationshipScore :: Maybe Double

The level of confidence that Amazon Comprehend Medical has that this attribute is correctly related to this entity.

beginOffset :: Maybe Int

The 0-based character offset in the input text that shows where the attribute begins. The offset returns the UTF-8 code point in the string.

text :: Maybe Text

The segment of input text extracted as this attribute.

category :: Maybe EntityType

The category of attribute.

score :: Maybe Double

The level of confidence that Amazon Comprehend Medical has that the segment of text is correctly recognized as an attribute.

traits :: Maybe [Trait]

Contextual information for this attribute.

relationshipType :: Maybe RelationshipType

The type of relationship between the entity and attribute. Type for the relationship is OVERLAP, indicating that the entity occurred at the same time as the Date_Expression.

endOffset :: Maybe Int

The 0-based character offset in the input text that shows where the attribute ends. The offset returns the UTF-8 code point in the string.

id :: Maybe Int

The numeric identifier for this attribute. This is a monotonically increasing id unique within this response rather than a global unique identifier.

type' :: Maybe EntitySubType

The type of attribute.

newAttribute :: Attribute

Create a value of Attribute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:relationshipScore:Attribute', attribute_relationshipScore - The level of confidence that Amazon Comprehend Medical has that this attribute is correctly related to this entity.

$sel:beginOffset:Attribute', attribute_beginOffset - The 0-based character offset in the input text that shows where the attribute begins. The offset returns the UTF-8 code point in the string.

$sel:text:Attribute', attribute_text - The segment of input text extracted as this attribute.

$sel:category:Attribute', attribute_category - The category of attribute.

$sel:score:Attribute', attribute_score - The level of confidence that Amazon Comprehend Medical has that the segment of text is correctly recognized as an attribute.

$sel:traits:Attribute', attribute_traits - Contextual information for this attribute.

$sel:relationshipType:Attribute', attribute_relationshipType - The type of relationship between the entity and attribute. Type for the relationship is OVERLAP, indicating that the entity occurred at the same time as the Date_Expression.

$sel:endOffset:Attribute', attribute_endOffset - The 0-based character offset in the input text that shows where the attribute ends. The offset returns the UTF-8 code point in the string.

$sel:id:Attribute', attribute_id - The numeric identifier for this attribute. This is a monotonically increasing id unique within this response rather than a global unique identifier.

$sel:type':Attribute', attribute_type - The type of attribute.

attribute_relationshipScore :: Lens' Attribute (Maybe Double)

The level of confidence that Amazon Comprehend Medical has that this attribute is correctly related to this entity.

attribute_beginOffset :: Lens' Attribute (Maybe Int)

The 0-based character offset in the input text that shows where the attribute begins. The offset returns the UTF-8 code point in the string.

attribute_text :: Lens' Attribute (Maybe Text)

The segment of input text extracted as this attribute.

attribute_category :: Lens' Attribute (Maybe EntityType)

The category of attribute.

attribute_score :: Lens' Attribute (Maybe Double)

The level of confidence that Amazon Comprehend Medical has that the segment of text is correctly recognized as an attribute.

attribute_traits :: Lens' Attribute (Maybe [Trait])

Contextual information for this attribute.

attribute_relationshipType :: Lens' Attribute (Maybe RelationshipType)

The type of relationship between the entity and attribute. Type for the relationship is OVERLAP, indicating that the entity occurred at the same time as the Date_Expression.

attribute_endOffset :: Lens' Attribute (Maybe Int)

The 0-based character offset in the input text that shows where the attribute ends. The offset returns the UTF-8 code point in the string.

attribute_id :: Lens' Attribute (Maybe Int)

The numeric identifier for this attribute. This is a monotonically increasing id unique within this response rather than a global unique identifier.

attribute_type :: Lens' Attribute (Maybe EntitySubType)

The type of attribute.