Network.AWS.ComprehendMedical.Types.RxNormAttribute

Description

Documentation

data RxNormAttribute

The extracted attributes that relate to this entity. The attributes recognized by InferRxNorm are DOSAGE, DURATION, FORM, FREQUENCY, RATE, ROUTE_OR_MODE.

See: newRxNormAttribute smart constructor.

Constructors

RxNormAttribute'

Fields

relationshipScore :: Maybe Double

The level of confidence that Amazon Comprehend Medical has that the attribute is accurately linked to an entity.

beginOffset :: Maybe Int

The 0-based character offset in the input text that shows where the attribute begins. The offset returns the UTF-8 code point in the string.

text :: Maybe Text

The segment of input text which corresponds to the detected attribute.

score :: Maybe Double

The level of confidence that Comprehend Medical has that the segment of text is correctly recognized as an attribute.

traits :: Maybe [RxNormTrait]

Contextual information for the attribute. InferRxNorm recognizes the trait NEGATION for attributes, i.e. that the patient is not taking a specific dose or form of a medication.

endOffset :: Maybe Int

The 0-based character offset in the input text that shows where the attribute ends. The offset returns the UTF-8 code point in the string.

id :: Maybe Int

The numeric identifier for this attribute. This is a monotonically increasing id unique within this response rather than a global unique identifier.

type' :: Maybe RxNormAttributeType

The type of attribute. The types of attributes recognized by InferRxNorm are BRAND_NAME and GENERIC_NAME.

newRxNormAttribute :: RxNormAttribute

Create a value of RxNormAttribute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:relationshipScore:RxNormAttribute', rxNormAttribute_relationshipScore - The level of confidence that Amazon Comprehend Medical has that the attribute is accurately linked to an entity.

$sel:beginOffset:RxNormAttribute', rxNormAttribute_beginOffset - The 0-based character offset in the input text that shows where the attribute begins. The offset returns the UTF-8 code point in the string.

$sel:text:RxNormAttribute', rxNormAttribute_text - The segment of input text which corresponds to the detected attribute.

$sel:score:RxNormAttribute', rxNormAttribute_score - The level of confidence that Comprehend Medical has that the segment of text is correctly recognized as an attribute.

$sel:traits:RxNormAttribute', rxNormAttribute_traits - Contextual information for the attribute. InferRxNorm recognizes the trait NEGATION for attributes, i.e. that the patient is not taking a specific dose or form of a medication.

$sel:endOffset:RxNormAttribute', rxNormAttribute_endOffset - The 0-based character offset in the input text that shows where the attribute ends. The offset returns the UTF-8 code point in the string.

$sel:id:RxNormAttribute', rxNormAttribute_id - The numeric identifier for this attribute. This is a monotonically increasing id unique within this response rather than a global unique identifier.

$sel:type':RxNormAttribute', rxNormAttribute_type - The type of attribute. The types of attributes recognized by InferRxNorm are BRAND_NAME and GENERIC_NAME.

rxNormAttribute_relationshipScore :: Lens' RxNormAttribute (Maybe Double)

The level of confidence that Amazon Comprehend Medical has that the attribute is accurately linked to an entity.

rxNormAttribute_beginOffset :: Lens' RxNormAttribute (Maybe Int)

The 0-based character offset in the input text that shows where the attribute begins. The offset returns the UTF-8 code point in the string.

rxNormAttribute_text :: Lens' RxNormAttribute (Maybe Text)

The segment of input text which corresponds to the detected attribute.

rxNormAttribute_score :: Lens' RxNormAttribute (Maybe Double)

The level of confidence that Comprehend Medical has that the segment of text is correctly recognized as an attribute.

rxNormAttribute_traits :: Lens' RxNormAttribute (Maybe [RxNormTrait])

Contextual information for the attribute. InferRxNorm recognizes the trait NEGATION for attributes, i.e. that the patient is not taking a specific dose or form of a medication.

rxNormAttribute_endOffset :: Lens' RxNormAttribute (Maybe Int)

The 0-based character offset in the input text that shows where the attribute ends. The offset returns the UTF-8 code point in the string.

rxNormAttribute_id :: Lens' RxNormAttribute (Maybe Int)

The numeric identifier for this attribute. This is a monotonically increasing id unique within this response rather than a global unique identifier.

rxNormAttribute_type :: Lens' RxNormAttribute (Maybe RxNormAttributeType)

The type of attribute. The types of attributes recognized by InferRxNorm are BRAND_NAME and GENERIC_NAME.