Network.AWS.ComprehendMedical.Types.RxNormTrait

Description

Documentation

data RxNormTrait

The contextual information for the entity. InferRxNorm recognizes the trait NEGATION, which is any indication that the patient is not taking a medication.

See: newRxNormTrait smart constructor.

Constructors

RxNormTrait'

Fields

score :: Maybe Double

The level of confidence that Amazon Comprehend Medical has in the accuracy of the detected trait.

name :: Maybe RxNormTraitName

Provides a name or contextual description about the trait.

newRxNormTrait :: RxNormTrait

Create a value of RxNormTrait with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:score:RxNormTrait', rxNormTrait_score - The level of confidence that Amazon Comprehend Medical has in the accuracy of the detected trait.

$sel:name:RxNormTrait', rxNormTrait_name - Provides a name or contextual description about the trait.

rxNormTrait_score :: Lens' RxNormTrait (Maybe Double)

The level of confidence that Amazon Comprehend Medical has in the accuracy of the detected trait.

rxNormTrait_name :: Lens' RxNormTrait (Maybe RxNormTraitName)

Provides a name or contextual description about the trait.