Network.AWS.ComprehendMedical.Types.UnmappedAttribute

Description

Documentation

data UnmappedAttribute

An attribute that we extracted, but were unable to relate to an entity.

See: newUnmappedAttribute smart constructor.

Constructors

UnmappedAttribute'

Fields

attribute :: Maybe Attribute

The specific attribute that has been extracted but not mapped to an entity.

type' :: Maybe EntityType

The type of the attribute, could be one of the following values: "MEDICATION", "MEDICAL_CONDITION", "ANATOMY", "TEST_AND_TREATMENT_PROCEDURE" or "PROTECTED_HEALTH_INFORMATION".

newUnmappedAttribute :: UnmappedAttribute

Create a value of UnmappedAttribute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attribute:UnmappedAttribute', unmappedAttribute_attribute - The specific attribute that has been extracted but not mapped to an entity.

$sel:type':UnmappedAttribute', unmappedAttribute_type - The type of the attribute, could be one of the following values: "MEDICATION", "MEDICAL_CONDITION", "ANATOMY", "TEST_AND_TREATMENT_PROCEDURE" or "PROTECTED_HEALTH_INFORMATION".

unmappedAttribute_attribute :: Lens' UnmappedAttribute (Maybe Attribute)

The specific attribute that has been extracted but not mapped to an entity.

unmappedAttribute_type :: Lens' UnmappedAttribute (Maybe EntityType)

The type of the attribute, could be one of the following values: "MEDICATION", "MEDICAL_CONDITION", "ANATOMY", "TEST_AND_TREATMENT_PROCEDURE" or "PROTECTED_HEALTH_INFORMATION".