Network.AWS.ComputeOptimizer.ExportEBSVolumeRecommendations

Description

Exports optimization recommendations for Amazon EBS volumes.

Recommendations are exported in a comma-separated values (.csv) file, and its metadata in a JavaScript Object Notation (JSON) (.json) file, to an existing Amazon Simple Storage Service (Amazon S3) bucket that you specify. For more information, see Exporting Recommendations in the Compute Optimizer User Guide.

You can have only one Amazon EBS volume export job in progress per Amazon Web Services Region.

Creating a Request

data ExportEBSVolumeRecommendations

See: newExportEBSVolumeRecommendations smart constructor.

Constructors

ExportEBSVolumeRecommendations'

Fields

accountIds :: Maybe [Text]

The IDs of the Amazon Web Services accounts for which to export Amazon EBS volume recommendations.

If your account is the management account of an organization, use this parameter to specify the member account for which you want to export recommendations.

This parameter cannot be specified together with the include member accounts parameter. The parameters are mutually exclusive.

Recommendations for member accounts are not included in the export if this parameter, or the include member accounts parameter, is omitted.

You can specify multiple account IDs per request.

fileFormat :: Maybe FileFormat

The format of the export file.

The only export file format currently supported is Csv.

filters :: Maybe [EBSFilter]

An array of objects to specify a filter that exports a more specific set of Amazon EBS volume recommendations.

fieldsToExport :: Maybe [ExportableVolumeField]

The recommendations data to include in the export file. For more information about the fields that can be exported, see Exported files in the Compute Optimizer User Guide.

includeMemberAccounts :: Maybe Bool

Indicates whether to include recommendations for resources in all member accounts of the organization if your account is the management account of an organization.

The member accounts must also be opted in to Compute Optimizer, and trusted access for Compute Optimizer must be enabled in the organization account. For more information, see Compute Optimizer and Amazon Web Services Organizations trusted access in the Compute Optimizer User Guide.

Recommendations for member accounts of the organization are not included in the export file if this parameter is omitted.

This parameter cannot be specified together with the account IDs parameter. The parameters are mutually exclusive.

Recommendations for member accounts are not included in the export if this parameter, or the account IDs parameter, is omitted.

s3DestinationConfig :: S3DestinationConfig

newExportEBSVolumeRecommendations

Arguments

:: S3DestinationConfig
$sel:s3DestinationConfig:ExportEBSVolumeRecommendations'

-> ExportEBSVolumeRecommendations

Create a value of ExportEBSVolumeRecommendations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountIds:ExportEBSVolumeRecommendations', exportEBSVolumeRecommendations_accountIds - The IDs of the Amazon Web Services accounts for which to export Amazon EBS volume recommendations.

If your account is the management account of an organization, use this parameter to specify the member account for which you want to export recommendations.

This parameter cannot be specified together with the include member accounts parameter. The parameters are mutually exclusive.

Recommendations for member accounts are not included in the export if this parameter, or the include member accounts parameter, is omitted.

You can specify multiple account IDs per request.

$sel:fileFormat:ExportEBSVolumeRecommendations', exportEBSVolumeRecommendations_fileFormat - The format of the export file.

The only export file format currently supported is Csv.

$sel:filters:ExportEBSVolumeRecommendations', exportEBSVolumeRecommendations_filters - An array of objects to specify a filter that exports a more specific set of Amazon EBS volume recommendations.

$sel:fieldsToExport:ExportEBSVolumeRecommendations', exportEBSVolumeRecommendations_fieldsToExport - The recommendations data to include in the export file. For more information about the fields that can be exported, see Exported files in the Compute Optimizer User Guide.

$sel:includeMemberAccounts:ExportEBSVolumeRecommendations', exportEBSVolumeRecommendations_includeMemberAccounts - Indicates whether to include recommendations for resources in all member accounts of the organization if your account is the management account of an organization.

The member accounts must also be opted in to Compute Optimizer, and trusted access for Compute Optimizer must be enabled in the organization account. For more information, see Compute Optimizer and Amazon Web Services Organizations trusted access in the Compute Optimizer User Guide.

Recommendations for member accounts of the organization are not included in the export file if this parameter is omitted.

This parameter cannot be specified together with the account IDs parameter. The parameters are mutually exclusive.

Recommendations for member accounts are not included in the export if this parameter, or the account IDs parameter, is omitted.

$sel:s3DestinationConfig:ExportEBSVolumeRecommendations', exportEBSVolumeRecommendations_s3DestinationConfig - Undocumented member.

Request Lenses

exportEBSVolumeRecommendations_accountIds :: Lens' ExportEBSVolumeRecommendations (Maybe [Text])

The IDs of the Amazon Web Services accounts for which to export Amazon EBS volume recommendations.

If your account is the management account of an organization, use this parameter to specify the member account for which you want to export recommendations.

This parameter cannot be specified together with the include member accounts parameter. The parameters are mutually exclusive.

Recommendations for member accounts are not included in the export if this parameter, or the include member accounts parameter, is omitted.

You can specify multiple account IDs per request.

exportEBSVolumeRecommendations_fileFormat :: Lens' ExportEBSVolumeRecommendations (Maybe FileFormat)

The format of the export file.

The only export file format currently supported is Csv.

exportEBSVolumeRecommendations_filters :: Lens' ExportEBSVolumeRecommendations (Maybe [EBSFilter])

An array of objects to specify a filter that exports a more specific set of Amazon EBS volume recommendations.

exportEBSVolumeRecommendations_fieldsToExport :: Lens' ExportEBSVolumeRecommendations (Maybe [ExportableVolumeField])

The recommendations data to include in the export file. For more information about the fields that can be exported, see Exported files in the Compute Optimizer User Guide.

exportEBSVolumeRecommendations_includeMemberAccounts :: Lens' ExportEBSVolumeRecommendations (Maybe Bool)

Indicates whether to include recommendations for resources in all member accounts of the organization if your account is the management account of an organization.

The member accounts must also be opted in to Compute Optimizer, and trusted access for Compute Optimizer must be enabled in the organization account. For more information, see Compute Optimizer and Amazon Web Services Organizations trusted access in the Compute Optimizer User Guide.

Recommendations for member accounts of the organization are not included in the export file if this parameter is omitted.

This parameter cannot be specified together with the account IDs parameter. The parameters are mutually exclusive.

Recommendations for member accounts are not included in the export if this parameter, or the account IDs parameter, is omitted.

exportEBSVolumeRecommendations_s3DestinationConfig :: Lens' ExportEBSVolumeRecommendations S3DestinationConfig

Undocumented member.

Destructuring the Response

data ExportEBSVolumeRecommendationsResponse

See: newExportEBSVolumeRecommendationsResponse smart constructor.

Constructors

ExportEBSVolumeRecommendationsResponse'

Fields

jobId :: Maybe Text

The identification number of the export job.

Use the DescribeRecommendationExportJobs action, and specify the job ID to view the status of an export job.

s3Destination :: Maybe S3Destination

httpStatus :: Int

The response's http status code.

newExportEBSVolumeRecommendationsResponse

Arguments

:: Int
$sel:httpStatus:ExportEBSVolumeRecommendationsResponse'

-> ExportEBSVolumeRecommendationsResponse

Create a value of ExportEBSVolumeRecommendationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobId:ExportEBSVolumeRecommendationsResponse', exportEBSVolumeRecommendationsResponse_jobId - The identification number of the export job.

Use the DescribeRecommendationExportJobs action, and specify the job ID to view the status of an export job.

$sel:s3Destination:ExportEBSVolumeRecommendationsResponse', exportEBSVolumeRecommendationsResponse_s3Destination - Undocumented member.

$sel:httpStatus:ExportEBSVolumeRecommendationsResponse', exportEBSVolumeRecommendationsResponse_httpStatus - The response's http status code.

Response Lenses

exportEBSVolumeRecommendationsResponse_jobId :: Lens' ExportEBSVolumeRecommendationsResponse (Maybe Text)

The identification number of the export job.

Use the DescribeRecommendationExportJobs action, and specify the job ID to view the status of an export job.

exportEBSVolumeRecommendationsResponse_s3Destination :: Lens' ExportEBSVolumeRecommendationsResponse (Maybe S3Destination)

Undocumented member.

exportEBSVolumeRecommendationsResponse_httpStatus :: Lens' ExportEBSVolumeRecommendationsResponse Int

The response's http status code.