Network.AWS.ComputeOptimizer.ExportEC2InstanceRecommendations

Description

Exports optimization recommendations for Amazon EC2 instances.

Recommendations are exported in a comma-separated values (.csv) file, and its metadata in a JavaScript Object Notation (JSON) (.json) file, to an existing Amazon Simple Storage Service (Amazon S3) bucket that you specify. For more information, see Exporting Recommendations in the Compute Optimizer User Guide.

You can have only one Amazon EC2 instance export job in progress per Amazon Web Services Region.

Creating a Request

data ExportEC2InstanceRecommendations

See: newExportEC2InstanceRecommendations smart constructor.

Constructors

ExportEC2InstanceRecommendations'

Fields

accountIds :: Maybe [Text]

The IDs of the Amazon Web Services accounts for which to export instance recommendations.

If your account is the management account of an organization, use this parameter to specify the member account for which you want to export recommendations.

This parameter cannot be specified together with the include member accounts parameter. The parameters are mutually exclusive.

Recommendations for member accounts are not included in the export if this parameter, or the include member accounts parameter, is omitted.

You can specify multiple account IDs per request.

fileFormat :: Maybe FileFormat

The format of the export file.

The only export file format currently supported is Csv.

filters :: Maybe [Filter]

An array of objects to specify a filter that exports a more specific set of instance recommendations.

fieldsToExport :: Maybe [ExportableInstanceField]

The recommendations data to include in the export file. For more information about the fields that can be exported, see Exported files in the Compute Optimizer User Guide.

includeMemberAccounts :: Maybe Bool

Indicates whether to include recommendations for resources in all member accounts of the organization if your account is the management account of an organization.

The member accounts must also be opted in to Compute Optimizer, and trusted access for Compute Optimizer must be enabled in the organization account. For more information, see Compute Optimizer and Amazon Web Services Organizations trusted access in the Compute Optimizer User Guide.

Recommendations for member accounts of the organization are not included in the export file if this parameter is omitted.

Recommendations for member accounts are not included in the export if this parameter, or the account IDs parameter, is omitted.

recommendationPreferences :: Maybe RecommendationPreferences

An object to specify the preferences for the Amazon EC2 instance recommendations to export.

s3DestinationConfig :: S3DestinationConfig

An object to specify the destination Amazon Simple Storage Service (Amazon S3) bucket name and key prefix for the export job.

You must create the destination Amazon S3 bucket for your recommendations export before you create the export job. Compute Optimizer does not create the S3 bucket for you. After you create the S3 bucket, ensure that it has the required permissions policy policy to allow Compute Optimizer to write the export file to it. If you plan to specify an object prefix when you create the export job, you must include the object prefix in the that you add to the S3 bucket. For more information, see Amazon S3 Bucket Policy for Compute Optimizer in the Compute Optimizer User Guide.

newExportEC2InstanceRecommendations

Arguments

:: S3DestinationConfig
$sel:s3DestinationConfig:ExportEC2InstanceRecommendations'

-> ExportEC2InstanceRecommendations

Create a value of ExportEC2InstanceRecommendations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountIds:ExportEC2InstanceRecommendations', exportEC2InstanceRecommendations_accountIds - The IDs of the Amazon Web Services accounts for which to export instance recommendations.

If your account is the management account of an organization, use this parameter to specify the member account for which you want to export recommendations.

This parameter cannot be specified together with the include member accounts parameter. The parameters are mutually exclusive.

Recommendations for member accounts are not included in the export if this parameter, or the include member accounts parameter, is omitted.

You can specify multiple account IDs per request.

$sel:fileFormat:ExportEC2InstanceRecommendations', exportEC2InstanceRecommendations_fileFormat - The format of the export file.

The only export file format currently supported is Csv.

$sel:filters:ExportEC2InstanceRecommendations', exportEC2InstanceRecommendations_filters - An array of objects to specify a filter that exports a more specific set of instance recommendations.

$sel:fieldsToExport:ExportEC2InstanceRecommendations', exportEC2InstanceRecommendations_fieldsToExport - The recommendations data to include in the export file. For more information about the fields that can be exported, see Exported files in the Compute Optimizer User Guide.

$sel:includeMemberAccounts:ExportEC2InstanceRecommendations', exportEC2InstanceRecommendations_includeMemberAccounts - Indicates whether to include recommendations for resources in all member accounts of the organization if your account is the management account of an organization.

The member accounts must also be opted in to Compute Optimizer, and trusted access for Compute Optimizer must be enabled in the organization account. For more information, see Compute Optimizer and Amazon Web Services Organizations trusted access in the Compute Optimizer User Guide.

Recommendations for member accounts of the organization are not included in the export file if this parameter is omitted.

Recommendations for member accounts are not included in the export if this parameter, or the account IDs parameter, is omitted.

$sel:recommendationPreferences:ExportEC2InstanceRecommendations', exportEC2InstanceRecommendations_recommendationPreferences - An object to specify the preferences for the Amazon EC2 instance recommendations to export.

$sel:s3DestinationConfig:ExportEC2InstanceRecommendations', exportEC2InstanceRecommendations_s3DestinationConfig - An object to specify the destination Amazon Simple Storage Service (Amazon S3) bucket name and key prefix for the export job.

You must create the destination Amazon S3 bucket for your recommendations export before you create the export job. Compute Optimizer does not create the S3 bucket for you. After you create the S3 bucket, ensure that it has the required permissions policy policy to allow Compute Optimizer to write the export file to it. If you plan to specify an object prefix when you create the export job, you must include the object prefix in the that you add to the S3 bucket. For more information, see Amazon S3 Bucket Policy for Compute Optimizer in the Compute Optimizer User Guide.

Request Lenses

exportEC2InstanceRecommendations_accountIds :: Lens' ExportEC2InstanceRecommendations (Maybe [Text])

The IDs of the Amazon Web Services accounts for which to export instance recommendations.

If your account is the management account of an organization, use this parameter to specify the member account for which you want to export recommendations.

This parameter cannot be specified together with the include member accounts parameter. The parameters are mutually exclusive.

Recommendations for member accounts are not included in the export if this parameter, or the include member accounts parameter, is omitted.

You can specify multiple account IDs per request.

exportEC2InstanceRecommendations_fileFormat :: Lens' ExportEC2InstanceRecommendations (Maybe FileFormat)

The format of the export file.

The only export file format currently supported is Csv.

exportEC2InstanceRecommendations_filters :: Lens' ExportEC2InstanceRecommendations (Maybe [Filter])

An array of objects to specify a filter that exports a more specific set of instance recommendations.

exportEC2InstanceRecommendations_fieldsToExport :: Lens' ExportEC2InstanceRecommendations (Maybe [ExportableInstanceField])

The recommendations data to include in the export file. For more information about the fields that can be exported, see Exported files in the Compute Optimizer User Guide.

exportEC2InstanceRecommendations_includeMemberAccounts :: Lens' ExportEC2InstanceRecommendations (Maybe Bool)

Indicates whether to include recommendations for resources in all member accounts of the organization if your account is the management account of an organization.

The member accounts must also be opted in to Compute Optimizer, and trusted access for Compute Optimizer must be enabled in the organization account. For more information, see Compute Optimizer and Amazon Web Services Organizations trusted access in the Compute Optimizer User Guide.

Recommendations for member accounts of the organization are not included in the export file if this parameter is omitted.

Recommendations for member accounts are not included in the export if this parameter, or the account IDs parameter, is omitted.

exportEC2InstanceRecommendations_recommendationPreferences :: Lens' ExportEC2InstanceRecommendations (Maybe RecommendationPreferences)

An object to specify the preferences for the Amazon EC2 instance recommendations to export.

exportEC2InstanceRecommendations_s3DestinationConfig :: Lens' ExportEC2InstanceRecommendations S3DestinationConfig

An object to specify the destination Amazon Simple Storage Service (Amazon S3) bucket name and key prefix for the export job.

You must create the destination Amazon S3 bucket for your recommendations export before you create the export job. Compute Optimizer does not create the S3 bucket for you. After you create the S3 bucket, ensure that it has the required permissions policy policy to allow Compute Optimizer to write the export file to it. If you plan to specify an object prefix when you create the export job, you must include the object prefix in the that you add to the S3 bucket. For more information, see Amazon S3 Bucket Policy for Compute Optimizer in the Compute Optimizer User Guide.

Destructuring the Response

data ExportEC2InstanceRecommendationsResponse

See: newExportEC2InstanceRecommendationsResponse smart constructor.

Constructors

ExportEC2InstanceRecommendationsResponse'

Fields

jobId :: Maybe Text

The identification number of the export job.

Use the DescribeRecommendationExportJobs action, and specify the job ID to view the status of an export job.

s3Destination :: Maybe S3Destination

An object that describes the destination Amazon S3 bucket of a recommendations export file.

httpStatus :: Int

The response's http status code.

newExportEC2InstanceRecommendationsResponse

Arguments

:: Int
$sel:httpStatus:ExportEC2InstanceRecommendationsResponse'

-> ExportEC2InstanceRecommendationsResponse

Create a value of ExportEC2InstanceRecommendationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobId:ExportEC2InstanceRecommendationsResponse', exportEC2InstanceRecommendationsResponse_jobId - The identification number of the export job.

Use the DescribeRecommendationExportJobs action, and specify the job ID to view the status of an export job.

$sel:s3Destination:ExportEC2InstanceRecommendationsResponse', exportEC2InstanceRecommendationsResponse_s3Destination - An object that describes the destination Amazon S3 bucket of a recommendations export file.

$sel:httpStatus:ExportEC2InstanceRecommendationsResponse', exportEC2InstanceRecommendationsResponse_httpStatus - The response's http status code.

Response Lenses

exportEC2InstanceRecommendationsResponse_jobId :: Lens' ExportEC2InstanceRecommendationsResponse (Maybe Text)

The identification number of the export job.

Use the DescribeRecommendationExportJobs action, and specify the job ID to view the status of an export job.

exportEC2InstanceRecommendationsResponse_s3Destination :: Lens' ExportEC2InstanceRecommendationsResponse (Maybe S3Destination)

An object that describes the destination Amazon S3 bucket of a recommendations export file.

exportEC2InstanceRecommendationsResponse_httpStatus :: Lens' ExportEC2InstanceRecommendationsResponse Int

The response's http status code.