Network.AWS.ComputeOptimizer.GetEC2InstanceRecommendations

Description

Returns Amazon EC2 instance recommendations.

Compute Optimizer generates recommendations for Amazon Elastic Compute Cloud (Amazon EC2) instances that meet a specific set of requirements. For more information, see the Supported resources and requirements in the Compute Optimizer User Guide.

Creating a Request

data GetEC2InstanceRecommendations

See: newGetEC2InstanceRecommendations smart constructor.

Constructors

GetEC2InstanceRecommendations'

Fields

accountIds :: Maybe [Text]

The ID of the Amazon Web Services account for which to return instance recommendations.

If your account is the management account of an organization, use this parameter to specify the member account for which you want to return instance recommendations.

Only one account ID can be specified per request.

filters :: Maybe [Filter]

An array of objects to specify a filter that returns a more specific list of instance recommendations.

recommendationPreferences :: Maybe RecommendationPreferences

An object to specify the preferences for the Amazon EC2 instance recommendations to return in the response.

nextToken :: Maybe Text

The token to advance to the next page of instance recommendations.

instanceArns :: Maybe [Text]

The Amazon Resource Name (ARN) of the instances for which to return recommendations.

maxResults :: Maybe Int

The maximum number of instance recommendations to return with a single request.

To retrieve the remaining results, make another request with the returned nextToken value.

newGetEC2InstanceRecommendations :: GetEC2InstanceRecommendations

Create a value of GetEC2InstanceRecommendations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountIds:GetEC2InstanceRecommendations', getEC2InstanceRecommendations_accountIds - The ID of the Amazon Web Services account for which to return instance recommendations.

If your account is the management account of an organization, use this parameter to specify the member account for which you want to return instance recommendations.

Only one account ID can be specified per request.

$sel:filters:GetEC2InstanceRecommendations', getEC2InstanceRecommendations_filters - An array of objects to specify a filter that returns a more specific list of instance recommendations.

$sel:recommendationPreferences:GetEC2InstanceRecommendations', getEC2InstanceRecommendations_recommendationPreferences - An object to specify the preferences for the Amazon EC2 instance recommendations to return in the response.

$sel:nextToken:GetEC2InstanceRecommendations', getEC2InstanceRecommendations_nextToken - The token to advance to the next page of instance recommendations.

$sel:instanceArns:GetEC2InstanceRecommendations', getEC2InstanceRecommendations_instanceArns - The Amazon Resource Name (ARN) of the instances for which to return recommendations.

$sel:maxResults:GetEC2InstanceRecommendations', getEC2InstanceRecommendations_maxResults - The maximum number of instance recommendations to return with a single request.

To retrieve the remaining results, make another request with the returned nextToken value.

Request Lenses

getEC2InstanceRecommendations_accountIds :: Lens' GetEC2InstanceRecommendations (Maybe [Text])

The ID of the Amazon Web Services account for which to return instance recommendations.

If your account is the management account of an organization, use this parameter to specify the member account for which you want to return instance recommendations.

Only one account ID can be specified per request.

getEC2InstanceRecommendations_filters :: Lens' GetEC2InstanceRecommendations (Maybe [Filter])

An array of objects to specify a filter that returns a more specific list of instance recommendations.

getEC2InstanceRecommendations_recommendationPreferences :: Lens' GetEC2InstanceRecommendations (Maybe RecommendationPreferences)

An object to specify the preferences for the Amazon EC2 instance recommendations to return in the response.

getEC2InstanceRecommendations_nextToken :: Lens' GetEC2InstanceRecommendations (Maybe Text)

The token to advance to the next page of instance recommendations.

getEC2InstanceRecommendations_instanceArns :: Lens' GetEC2InstanceRecommendations (Maybe [Text])

The Amazon Resource Name (ARN) of the instances for which to return recommendations.

getEC2InstanceRecommendations_maxResults :: Lens' GetEC2InstanceRecommendations (Maybe Int)

The maximum number of instance recommendations to return with a single request.

To retrieve the remaining results, make another request with the returned nextToken value.

Destructuring the Response

data GetEC2InstanceRecommendationsResponse

See: newGetEC2InstanceRecommendationsResponse smart constructor.

Constructors

GetEC2InstanceRecommendationsResponse'

Fields

nextToken :: Maybe Text

The token to use to advance to the next page of instance recommendations.

This value is null when there are no more pages of instance recommendations to return.

errors :: Maybe [GetRecommendationError]

An array of objects that describe errors of the request.

For example, an error is returned if you request recommendations for an instance of an unsupported instance family.

instanceRecommendations :: Maybe [InstanceRecommendation]

An array of objects that describe instance recommendations.

httpStatus :: Int

The response's http status code.

newGetEC2InstanceRecommendationsResponse

Arguments

:: Int
$sel:httpStatus:GetEC2InstanceRecommendationsResponse'

-> GetEC2InstanceRecommendationsResponse

Create a value of GetEC2InstanceRecommendationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetEC2InstanceRecommendations', getEC2InstanceRecommendationsResponse_nextToken - The token to use to advance to the next page of instance recommendations.

This value is null when there are no more pages of instance recommendations to return.

$sel:errors:GetEC2InstanceRecommendationsResponse', getEC2InstanceRecommendationsResponse_errors - An array of objects that describe errors of the request.

For example, an error is returned if you request recommendations for an instance of an unsupported instance family.

$sel:instanceRecommendations:GetEC2InstanceRecommendationsResponse', getEC2InstanceRecommendationsResponse_instanceRecommendations - An array of objects that describe instance recommendations.

$sel:httpStatus:GetEC2InstanceRecommendationsResponse', getEC2InstanceRecommendationsResponse_httpStatus - The response's http status code.

Response Lenses

getEC2InstanceRecommendationsResponse_nextToken :: Lens' GetEC2InstanceRecommendationsResponse (Maybe Text)

The token to use to advance to the next page of instance recommendations.

This value is null when there are no more pages of instance recommendations to return.

getEC2InstanceRecommendationsResponse_errors :: Lens' GetEC2InstanceRecommendationsResponse (Maybe [GetRecommendationError])

An array of objects that describe errors of the request.

For example, an error is returned if you request recommendations for an instance of an unsupported instance family.

getEC2InstanceRecommendationsResponse_instanceRecommendations :: Lens' GetEC2InstanceRecommendationsResponse (Maybe [InstanceRecommendation])

An array of objects that describe instance recommendations.

getEC2InstanceRecommendationsResponse_httpStatus :: Lens' GetEC2InstanceRecommendationsResponse Int

The response's http status code.