Network.AWS.ComputeOptimizer.Types.AccountEnrollmentStatus

Description

Documentation

data AccountEnrollmentStatus

Describes the enrollment status of an organization's member accounts in Compute Optimizer.

See: newAccountEnrollmentStatus smart constructor.

Constructors

AccountEnrollmentStatus'

Fields

status :: Maybe Status

The account enrollment status.

accountId :: Maybe Text

The Amazon Web Services account ID.

statusReason :: Maybe Text

The reason for the account enrollment status.

For example, an account might show a status of Pending because member accounts of an organization require more time to be enrolled in the service.

lastUpdatedTimestamp :: Maybe POSIX

The Unix epoch timestamp, in seconds, of when the account enrollment status was last updated.

newAccountEnrollmentStatus :: AccountEnrollmentStatus

Create a value of AccountEnrollmentStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:AccountEnrollmentStatus', accountEnrollmentStatus_status - The account enrollment status.

$sel:accountId:AccountEnrollmentStatus', accountEnrollmentStatus_accountId - The Amazon Web Services account ID.

$sel:statusReason:AccountEnrollmentStatus', accountEnrollmentStatus_statusReason - The reason for the account enrollment status.

For example, an account might show a status of Pending because member accounts of an organization require more time to be enrolled in the service.

$sel:lastUpdatedTimestamp:AccountEnrollmentStatus', accountEnrollmentStatus_lastUpdatedTimestamp - The Unix epoch timestamp, in seconds, of when the account enrollment status was last updated.

accountEnrollmentStatus_status :: Lens' AccountEnrollmentStatus (Maybe Status)

The account enrollment status.

accountEnrollmentStatus_accountId :: Lens' AccountEnrollmentStatus (Maybe Text)

The Amazon Web Services account ID.

accountEnrollmentStatus_statusReason :: Lens' AccountEnrollmentStatus (Maybe Text)

The reason for the account enrollment status.

For example, an account might show a status of Pending because member accounts of an organization require more time to be enrolled in the service.

accountEnrollmentStatus_lastUpdatedTimestamp :: Lens' AccountEnrollmentStatus (Maybe UTCTime)

The Unix epoch timestamp, in seconds, of when the account enrollment status was last updated.