Network.AWS.ComputeOptimizer.Types.AutoScalingGroupRecommendation

Description

Documentation

data AutoScalingGroupRecommendation

Describes an Auto Scaling group recommendation.

See: newAutoScalingGroupRecommendation smart constructor.

Constructors

AutoScalingGroupRecommendation'

Fields

finding :: Maybe Finding

The finding classification of the Auto Scaling group.

Findings for Auto Scaling groups include:

NotOptimized —An Auto Scaling group is considered not optimized when Compute Optimizer identifies a recommendation that can provide better performance for your workload.

Optimized —An Auto Scaling group is considered optimized when Compute Optimizer determines that the group is correctly provisioned to run your workload based on the chosen instance type. For optimized resources, Compute Optimizer might recommend a new generation instance type.

lastRefreshTimestamp :: Maybe POSIX

The timestamp of when the Auto Scaling group recommendation was last refreshed.

currentConfiguration :: Maybe AutoScalingGroupConfiguration

An array of objects that describe the current configuration of the Auto Scaling group.

accountId :: Maybe Text

The Amazon Web Services account ID of the Auto Scaling group.

autoScalingGroupName :: Maybe Text

The name of the Auto Scaling group.

utilizationMetrics :: Maybe [UtilizationMetric]

An array of objects that describe the utilization metrics of the Auto Scaling group.

autoScalingGroupArn :: Maybe Text

The Amazon Resource Name (ARN) of the Auto Scaling group.

recommendationOptions :: Maybe [AutoScalingGroupRecommendationOption]

An array of objects that describe the recommendation options for the Auto Scaling group.

lookBackPeriodInDays :: Maybe Double

The number of days for which utilization metrics were analyzed for the Auto Scaling group.

newAutoScalingGroupRecommendation :: AutoScalingGroupRecommendation

Create a value of AutoScalingGroupRecommendation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:finding:AutoScalingGroupRecommendation', autoScalingGroupRecommendation_finding - The finding classification of the Auto Scaling group.

Findings for Auto Scaling groups include:

NotOptimized —An Auto Scaling group is considered not optimized when Compute Optimizer identifies a recommendation that can provide better performance for your workload.

Optimized —An Auto Scaling group is considered optimized when Compute Optimizer determines that the group is correctly provisioned to run your workload based on the chosen instance type. For optimized resources, Compute Optimizer might recommend a new generation instance type.

$sel:lastRefreshTimestamp:AutoScalingGroupRecommendation', autoScalingGroupRecommendation_lastRefreshTimestamp - The timestamp of when the Auto Scaling group recommendation was last refreshed.

$sel:currentConfiguration:AutoScalingGroupRecommendation', autoScalingGroupRecommendation_currentConfiguration - An array of objects that describe the current configuration of the Auto Scaling group.

$sel:accountId:AutoScalingGroupRecommendation', autoScalingGroupRecommendation_accountId - The Amazon Web Services account ID of the Auto Scaling group.

$sel:autoScalingGroupName:AutoScalingGroupRecommendation', autoScalingGroupRecommendation_autoScalingGroupName - The name of the Auto Scaling group.

$sel:utilizationMetrics:AutoScalingGroupRecommendation', autoScalingGroupRecommendation_utilizationMetrics - An array of objects that describe the utilization metrics of the Auto Scaling group.

$sel:autoScalingGroupArn:AutoScalingGroupRecommendation', autoScalingGroupRecommendation_autoScalingGroupArn - The Amazon Resource Name (ARN) of the Auto Scaling group.

$sel:recommendationOptions:AutoScalingGroupRecommendation', autoScalingGroupRecommendation_recommendationOptions - An array of objects that describe the recommendation options for the Auto Scaling group.

$sel:lookBackPeriodInDays:AutoScalingGroupRecommendation', autoScalingGroupRecommendation_lookBackPeriodInDays - The number of days for which utilization metrics were analyzed for the Auto Scaling group.

autoScalingGroupRecommendation_finding :: Lens' AutoScalingGroupRecommendation (Maybe Finding)

The finding classification of the Auto Scaling group.

Findings for Auto Scaling groups include:

NotOptimized —An Auto Scaling group is considered not optimized when Compute Optimizer identifies a recommendation that can provide better performance for your workload.

Optimized —An Auto Scaling group is considered optimized when Compute Optimizer determines that the group is correctly provisioned to run your workload based on the chosen instance type. For optimized resources, Compute Optimizer might recommend a new generation instance type.

autoScalingGroupRecommendation_lastRefreshTimestamp :: Lens' AutoScalingGroupRecommendation (Maybe UTCTime)

The timestamp of when the Auto Scaling group recommendation was last refreshed.

autoScalingGroupRecommendation_currentConfiguration :: Lens' AutoScalingGroupRecommendation (Maybe AutoScalingGroupConfiguration)

An array of objects that describe the current configuration of the Auto Scaling group.

autoScalingGroupRecommendation_accountId :: Lens' AutoScalingGroupRecommendation (Maybe Text)

The Amazon Web Services account ID of the Auto Scaling group.

autoScalingGroupRecommendation_autoScalingGroupName :: Lens' AutoScalingGroupRecommendation (Maybe Text)

The name of the Auto Scaling group.

autoScalingGroupRecommendation_utilizationMetrics :: Lens' AutoScalingGroupRecommendation (Maybe [UtilizationMetric])

An array of objects that describe the utilization metrics of the Auto Scaling group.

autoScalingGroupRecommendation_autoScalingGroupArn :: Lens' AutoScalingGroupRecommendation (Maybe Text)

The Amazon Resource Name (ARN) of the Auto Scaling group.

autoScalingGroupRecommendation_recommendationOptions :: Lens' AutoScalingGroupRecommendation (Maybe [AutoScalingGroupRecommendationOption])

An array of objects that describe the recommendation options for the Auto Scaling group.

autoScalingGroupRecommendation_lookBackPeriodInDays :: Lens' AutoScalingGroupRecommendation (Maybe Double)

The number of days for which utilization metrics were analyzed for the Auto Scaling group.