Network.AWS.ComputeOptimizer.Types.AutoScalingGroupRecommendationOption

Description

Documentation

data AutoScalingGroupRecommendationOption

Describes a recommendation option for an Auto Scaling group.

See: newAutoScalingGroupRecommendationOption smart constructor.

Constructors

AutoScalingGroupRecommendationOption'

Fields

performanceRisk :: Maybe Double

The performance risk of the Auto Scaling group configuration recommendation.

Performance risk indicates the likelihood of the recommended instance type not meeting the resource needs of your workload. Compute Optimizer calculates an individual performance risk score for each specification of the recommended instance, including CPU, memory, EBS throughput, EBS IOPS, disk throughput, disk IOPS, network throughput, and network PPS. The performance risk of the recommended instance is calculated as the maximum performance risk score across the analyzed resource specifications.

The value ranges from 0 - 4, with 0 meaning that the recommended resource is predicted to always provide enough hardware capability. The higher the performance risk is, the more likely you should validate whether the recommendation will meet the performance requirements of your workload before migrating your resource.

projectedUtilizationMetrics :: Maybe [UtilizationMetric]

An array of objects that describe the projected utilization metrics of the Auto Scaling group recommendation option.

The Cpu and Memory metrics are the only projected utilization metrics returned. Additionally, the Memory metric is returned only for resources that have the unified CloudWatch agent installed on them. For more information, see Enabling Memory Utilization with the CloudWatch Agent.

configuration :: Maybe AutoScalingGroupConfiguration

An array of objects that describe an Auto Scaling group configuration.

rank :: Maybe Int

The rank of the Auto Scaling group recommendation option.

The top recommendation option is ranked as 1.

newAutoScalingGroupRecommendationOption :: AutoScalingGroupRecommendationOption

Create a value of AutoScalingGroupRecommendationOption with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:performanceRisk:AutoScalingGroupRecommendationOption', autoScalingGroupRecommendationOption_performanceRisk - The performance risk of the Auto Scaling group configuration recommendation.

Performance risk indicates the likelihood of the recommended instance type not meeting the resource needs of your workload. Compute Optimizer calculates an individual performance risk score for each specification of the recommended instance, including CPU, memory, EBS throughput, EBS IOPS, disk throughput, disk IOPS, network throughput, and network PPS. The performance risk of the recommended instance is calculated as the maximum performance risk score across the analyzed resource specifications.

The value ranges from 0 - 4, with 0 meaning that the recommended resource is predicted to always provide enough hardware capability. The higher the performance risk is, the more likely you should validate whether the recommendation will meet the performance requirements of your workload before migrating your resource.

$sel:projectedUtilizationMetrics:AutoScalingGroupRecommendationOption', autoScalingGroupRecommendationOption_projectedUtilizationMetrics - An array of objects that describe the projected utilization metrics of the Auto Scaling group recommendation option.

The Cpu and Memory metrics are the only projected utilization metrics returned. Additionally, the Memory metric is returned only for resources that have the unified CloudWatch agent installed on them. For more information, see Enabling Memory Utilization with the CloudWatch Agent.

$sel:configuration:AutoScalingGroupRecommendationOption', autoScalingGroupRecommendationOption_configuration - An array of objects that describe an Auto Scaling group configuration.

$sel:rank:AutoScalingGroupRecommendationOption', autoScalingGroupRecommendationOption_rank - The rank of the Auto Scaling group recommendation option.

The top recommendation option is ranked as 1.

autoScalingGroupRecommendationOption_performanceRisk :: Lens' AutoScalingGroupRecommendationOption (Maybe Double)

The performance risk of the Auto Scaling group configuration recommendation.

Performance risk indicates the likelihood of the recommended instance type not meeting the resource needs of your workload. Compute Optimizer calculates an individual performance risk score for each specification of the recommended instance, including CPU, memory, EBS throughput, EBS IOPS, disk throughput, disk IOPS, network throughput, and network PPS. The performance risk of the recommended instance is calculated as the maximum performance risk score across the analyzed resource specifications.

The value ranges from 0 - 4, with 0 meaning that the recommended resource is predicted to always provide enough hardware capability. The higher the performance risk is, the more likely you should validate whether the recommendation will meet the performance requirements of your workload before migrating your resource.

autoScalingGroupRecommendationOption_projectedUtilizationMetrics :: Lens' AutoScalingGroupRecommendationOption (Maybe [UtilizationMetric])

An array of objects that describe the projected utilization metrics of the Auto Scaling group recommendation option.

The Cpu and Memory metrics are the only projected utilization metrics returned. Additionally, the Memory metric is returned only for resources that have the unified CloudWatch agent installed on them. For more information, see Enabling Memory Utilization with the CloudWatch Agent.

autoScalingGroupRecommendationOption_configuration :: Lens' AutoScalingGroupRecommendationOption (Maybe AutoScalingGroupConfiguration)

An array of objects that describe an Auto Scaling group configuration.

autoScalingGroupRecommendationOption_rank :: Lens' AutoScalingGroupRecommendationOption (Maybe Int)

The rank of the Auto Scaling group recommendation option.

The top recommendation option is ranked as 1.