Network.AWS.ComputeOptimizer.Types.EBSUtilizationMetric

Description

Documentation

data EBSUtilizationMetric

Describes a utilization metric of an Amazon Elastic Block Store (Amazon EBS) volume.

Compare the utilization metric data of your resource against its projected utilization metric data to determine the performance difference between your current resource and the recommended option.

See: newEBSUtilizationMetric smart constructor.

Constructors

EBSUtilizationMetric'

Fields

value :: Maybe Double

The value of the utilization metric.

name :: Maybe EBSMetricName

The name of the utilization metric.

The following utilization metrics are available:

VolumeReadOpsPerSecond - The completed read operations per second from the volume in a specified period of time.

Unit: Count

VolumeWriteOpsPerSecond - The completed write operations per second to the volume in a specified period of time.

Unit: Count

VolumeReadBytesPerSecond - The bytes read per second from the volume in a specified period of time.

Unit: Bytes

VolumeWriteBytesPerSecond - The bytes written to the volume in a specified period of time.

Unit: Bytes

statistic :: Maybe MetricStatistic

The statistic of the utilization metric.

The Compute Optimizer API, Command Line Interface (CLI), and SDKs return utilization metrics using only the Maximum statistic, which is the highest value observed during the specified period.

The Compute Optimizer console displays graphs for some utilization metrics using the Average statistic, which is the value of Sum / SampleCount during the specified period. For more information, see Viewing resource recommendations in the Compute Optimizer User Guide. You can also get averaged utilization metric data for your resources using Amazon CloudWatch. For more information, see the Amazon CloudWatch User Guide.

newEBSUtilizationMetric :: EBSUtilizationMetric

Create a value of EBSUtilizationMetric with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:EBSUtilizationMetric', eBSUtilizationMetric_value - The value of the utilization metric.

$sel:name:EBSUtilizationMetric', eBSUtilizationMetric_name - The name of the utilization metric.

The following utilization metrics are available:

VolumeReadOpsPerSecond - The completed read operations per second from the volume in a specified period of time.

Unit: Count

VolumeWriteOpsPerSecond - The completed write operations per second to the volume in a specified period of time.

Unit: Count

VolumeReadBytesPerSecond - The bytes read per second from the volume in a specified period of time.

Unit: Bytes

VolumeWriteBytesPerSecond - The bytes written to the volume in a specified period of time.

Unit: Bytes

$sel:statistic:EBSUtilizationMetric', eBSUtilizationMetric_statistic - The statistic of the utilization metric.

The Compute Optimizer API, Command Line Interface (CLI), and SDKs return utilization metrics using only the Maximum statistic, which is the highest value observed during the specified period.

The Compute Optimizer console displays graphs for some utilization metrics using the Average statistic, which is the value of Sum / SampleCount during the specified period. For more information, see Viewing resource recommendations in the Compute Optimizer User Guide. You can also get averaged utilization metric data for your resources using Amazon CloudWatch. For more information, see the Amazon CloudWatch User Guide.

eBSUtilizationMetric_value :: Lens' EBSUtilizationMetric (Maybe Double)

The value of the utilization metric.

eBSUtilizationMetric_name :: Lens' EBSUtilizationMetric (Maybe EBSMetricName)

The name of the utilization metric.

The following utilization metrics are available:

VolumeReadOpsPerSecond - The completed read operations per second from the volume in a specified period of time.

Unit: Count

VolumeWriteOpsPerSecond - The completed write operations per second to the volume in a specified period of time.

Unit: Count

VolumeReadBytesPerSecond - The bytes read per second from the volume in a specified period of time.

Unit: Bytes

VolumeWriteBytesPerSecond - The bytes written to the volume in a specified period of time.

Unit: Bytes

eBSUtilizationMetric_statistic :: Lens' EBSUtilizationMetric (Maybe MetricStatistic)

The statistic of the utilization metric.

The Compute Optimizer API, Command Line Interface (CLI), and SDKs return utilization metrics using only the Maximum statistic, which is the highest value observed during the specified period.

The Compute Optimizer console displays graphs for some utilization metrics using the Average statistic, which is the value of Sum / SampleCount during the specified period. For more information, see Viewing resource recommendations in the Compute Optimizer User Guide. You can also get averaged utilization metric data for your resources using Amazon CloudWatch. For more information, see the Amazon CloudWatch User Guide.