Network.AWS.ComputeOptimizer.Types.EnrollmentFilter

Description

Documentation

data EnrollmentFilter

Describes a filter that returns a more specific list of account enrollment statuses. Use this filter with the GetEnrollmentStatusesForOrganization action.

See: newEnrollmentFilter smart constructor.

Constructors

EnrollmentFilter'

Fields

values :: Maybe [Text]

The value of the filter.

The valid values are Active, Inactive, Pending, and Failed.

name :: Maybe EnrollmentFilterName

The name of the filter.

Specify Status to return accounts with a specific enrollment status (for example, Active).

newEnrollmentFilter :: EnrollmentFilter

Create a value of EnrollmentFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:values:EnrollmentFilter', enrollmentFilter_values - The value of the filter.

The valid values are Active, Inactive, Pending, and Failed.

$sel:name:EnrollmentFilter', enrollmentFilter_name - The name of the filter.

Specify Status to return accounts with a specific enrollment status (for example, Active).

enrollmentFilter_values :: Lens' EnrollmentFilter (Maybe [Text])

The value of the filter.

The valid values are Active, Inactive, Pending, and Failed.

enrollmentFilter_name :: Lens' EnrollmentFilter (Maybe EnrollmentFilterName)

The name of the filter.

Specify Status to return accounts with a specific enrollment status (for example, Active).