Network.AWS.ComputeOptimizer.Types.RecommendationExportJob

Description

Documentation

data RecommendationExportJob

Describes a recommendation export job.

Use the DescribeRecommendationExportJobs action to view your recommendation export jobs.

Use the ExportAutoScalingGroupRecommendations or ExportEC2InstanceRecommendations actions to request an export of your recommendations.

See: newRecommendationExportJob smart constructor.

Constructors

RecommendationExportJob'

Fields

failureReason :: Maybe Text

The reason for an export job failure.

destination :: Maybe ExportDestination

An object that describes the destination of the export file.

status :: Maybe JobStatus

The status of the export job.

jobId :: Maybe Text

The identification number of the export job.

resourceType :: Maybe ResourceType

The resource type of the exported recommendations.

creationTimestamp :: Maybe POSIX

The timestamp of when the export job was created.

lastUpdatedTimestamp :: Maybe POSIX

The timestamp of when the export job was last updated.

newRecommendationExportJob :: RecommendationExportJob

Create a value of RecommendationExportJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:RecommendationExportJob', recommendationExportJob_failureReason - The reason for an export job failure.

$sel:destination:RecommendationExportJob', recommendationExportJob_destination - An object that describes the destination of the export file.

$sel:status:RecommendationExportJob', recommendationExportJob_status - The status of the export job.

$sel:jobId:RecommendationExportJob', recommendationExportJob_jobId - The identification number of the export job.

$sel:resourceType:RecommendationExportJob', recommendationExportJob_resourceType - The resource type of the exported recommendations.

$sel:creationTimestamp:RecommendationExportJob', recommendationExportJob_creationTimestamp - The timestamp of when the export job was created.

$sel:lastUpdatedTimestamp:RecommendationExportJob', recommendationExportJob_lastUpdatedTimestamp - The timestamp of when the export job was last updated.

recommendationExportJob_failureReason :: Lens' RecommendationExportJob (Maybe Text)

The reason for an export job failure.

recommendationExportJob_destination :: Lens' RecommendationExportJob (Maybe ExportDestination)

An object that describes the destination of the export file.

recommendationExportJob_status :: Lens' RecommendationExportJob (Maybe JobStatus)

The status of the export job.

recommendationExportJob_jobId :: Lens' RecommendationExportJob (Maybe Text)

The identification number of the export job.

recommendationExportJob_resourceType :: Lens' RecommendationExportJob (Maybe ResourceType)

The resource type of the exported recommendations.

recommendationExportJob_creationTimestamp :: Lens' RecommendationExportJob (Maybe UTCTime)

The timestamp of when the export job was created.

recommendationExportJob_lastUpdatedTimestamp :: Lens' RecommendationExportJob (Maybe UTCTime)

The timestamp of when the export job was last updated.