Network.AWS.ComputeOptimizer.Types.RecommendedOptionProjectedMetric

Description

Documentation

data RecommendedOptionProjectedMetric

Describes a projected utilization metric of a recommendation option.

The Cpu and Memory metrics are the only projected utilization metrics returned when you run the GetEC2RecommendationProjectedMetrics action. Additionally, the Memory metric is returned only for resources that have the unified CloudWatch agent installed on them. For more information, see Enabling Memory Utilization with the CloudWatch Agent.

See: newRecommendedOptionProjectedMetric smart constructor.

Constructors

RecommendedOptionProjectedMetric'

Fields

recommendedInstanceType :: Maybe Text

The recommended instance type.

projectedMetrics :: Maybe [ProjectedMetric]

An array of objects that describe a projected utilization metric.

rank :: Maybe Int

The rank of the recommendation option projected metric.

The top recommendation option is ranked as 1.

The projected metric rank correlates to the recommendation option rank. For example, the projected metric ranked as 1 is related to the recommendation option that is also ranked as 1 in the same response.

newRecommendedOptionProjectedMetric :: RecommendedOptionProjectedMetric

Create a value of RecommendedOptionProjectedMetric with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:recommendedInstanceType:RecommendedOptionProjectedMetric', recommendedOptionProjectedMetric_recommendedInstanceType - The recommended instance type.

$sel:projectedMetrics:RecommendedOptionProjectedMetric', recommendedOptionProjectedMetric_projectedMetrics - An array of objects that describe a projected utilization metric.

$sel:rank:RecommendedOptionProjectedMetric', recommendedOptionProjectedMetric_rank - The rank of the recommendation option projected metric.

The top recommendation option is ranked as 1.

The projected metric rank correlates to the recommendation option rank. For example, the projected metric ranked as 1 is related to the recommendation option that is also ranked as 1 in the same response.

recommendedOptionProjectedMetric_recommendedInstanceType :: Lens' RecommendedOptionProjectedMetric (Maybe Text)

The recommended instance type.

recommendedOptionProjectedMetric_projectedMetrics :: Lens' RecommendedOptionProjectedMetric (Maybe [ProjectedMetric])

An array of objects that describe a projected utilization metric.

recommendedOptionProjectedMetric_rank :: Lens' RecommendedOptionProjectedMetric (Maybe Int)

The rank of the recommendation option projected metric.

The top recommendation option is ranked as 1.

The projected metric rank correlates to the recommendation option rank. For example, the projected metric ranked as 1 is related to the recommendation option that is also ranked as 1 in the same response.