Network.AWS.ComputeOptimizer.Types.S3DestinationConfig

Description

Documentation

data S3DestinationConfig

Describes the destination Amazon Simple Storage Service (Amazon S3) bucket name and key prefix for a recommendations export job.

You must create the destination Amazon S3 bucket for your recommendations export before you create the export job. Compute Optimizer does not create the S3 bucket for you. After you create the S3 bucket, ensure that it has the required permission policy to allow Compute Optimizer to write the export file to it. If you plan to specify an object prefix when you create the export job, you must include the object prefix in the policy that you add to the S3 bucket. For more information, see Amazon S3 Bucket Policy for Compute Optimizer in the Compute Optimizer User Guide.

See: newS3DestinationConfig smart constructor.

Constructors

S3DestinationConfig'

Fields

bucket :: Maybe Text

The name of the Amazon S3 bucket to use as the destination for an export job.

keyPrefix :: Maybe Text

The Amazon S3 bucket prefix for an export job.

newS3DestinationConfig :: S3DestinationConfig

Create a value of S3DestinationConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bucket:S3DestinationConfig', s3DestinationConfig_bucket - The name of the Amazon S3 bucket to use as the destination for an export job.

$sel:keyPrefix:S3DestinationConfig', s3DestinationConfig_keyPrefix - The Amazon S3 bucket prefix for an export job.

s3DestinationConfig_bucket :: Lens' S3DestinationConfig (Maybe Text)

The name of the Amazon S3 bucket to use as the destination for an export job.

s3DestinationConfig_keyPrefix :: Lens' S3DestinationConfig (Maybe Text)

The Amazon S3 bucket prefix for an export job.