Network.AWS.ComputeOptimizer.Types.VolumeRecommendation

Description

Documentation

data VolumeRecommendation

Describes an Amazon Elastic Block Store (Amazon EBS) volume recommendation.

See: newVolumeRecommendation smart constructor.

Constructors

VolumeRecommendation'

Fields

finding :: Maybe EBSFinding

The finding classification of the volume.

Findings for volumes include:

NotOptimized —A volume is considered not optimized when Compute Optimizer identifies a recommendation that can provide better performance for your workload.

Optimized —An volume is considered optimized when Compute Optimizer determines that the volume is correctly provisioned to run your workload based on the chosen volume type. For optimized resources, Compute Optimizer might recommend a new generation volume type.

volumeArn :: Maybe Text

The Amazon Resource Name (ARN) of the current volume.

lastRefreshTimestamp :: Maybe POSIX

The timestamp of when the volume recommendation was last refreshed.

currentConfiguration :: Maybe VolumeConfiguration

An array of objects that describe the current configuration of the volume.

accountId :: Maybe Text

The Amazon Web Services account ID of the volume.

utilizationMetrics :: Maybe [EBSUtilizationMetric]

An array of objects that describe the utilization metrics of the volume.

volumeRecommendationOptions :: Maybe [VolumeRecommendationOption]

An array of objects that describe the recommendation options for the volume.

lookBackPeriodInDays :: Maybe Double

The number of days for which utilization metrics were analyzed for the volume.

newVolumeRecommendation :: VolumeRecommendation

Create a value of VolumeRecommendation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:finding:VolumeRecommendation', volumeRecommendation_finding - The finding classification of the volume.

Findings for volumes include:

NotOptimized —A volume is considered not optimized when Compute Optimizer identifies a recommendation that can provide better performance for your workload.

Optimized —An volume is considered optimized when Compute Optimizer determines that the volume is correctly provisioned to run your workload based on the chosen volume type. For optimized resources, Compute Optimizer might recommend a new generation volume type.

$sel:volumeArn:VolumeRecommendation', volumeRecommendation_volumeArn - The Amazon Resource Name (ARN) of the current volume.

$sel:lastRefreshTimestamp:VolumeRecommendation', volumeRecommendation_lastRefreshTimestamp - The timestamp of when the volume recommendation was last refreshed.

$sel:currentConfiguration:VolumeRecommendation', volumeRecommendation_currentConfiguration - An array of objects that describe the current configuration of the volume.

$sel:accountId:VolumeRecommendation', volumeRecommendation_accountId - The Amazon Web Services account ID of the volume.

$sel:utilizationMetrics:VolumeRecommendation', volumeRecommendation_utilizationMetrics - An array of objects that describe the utilization metrics of the volume.

$sel:volumeRecommendationOptions:VolumeRecommendation', volumeRecommendation_volumeRecommendationOptions - An array of objects that describe the recommendation options for the volume.

$sel:lookBackPeriodInDays:VolumeRecommendation', volumeRecommendation_lookBackPeriodInDays - The number of days for which utilization metrics were analyzed for the volume.

volumeRecommendation_finding :: Lens' VolumeRecommendation (Maybe EBSFinding)

The finding classification of the volume.

Findings for volumes include:

NotOptimized —A volume is considered not optimized when Compute Optimizer identifies a recommendation that can provide better performance for your workload.

Optimized —An volume is considered optimized when Compute Optimizer determines that the volume is correctly provisioned to run your workload based on the chosen volume type. For optimized resources, Compute Optimizer might recommend a new generation volume type.

volumeRecommendation_volumeArn :: Lens' VolumeRecommendation (Maybe Text)

The Amazon Resource Name (ARN) of the current volume.

volumeRecommendation_lastRefreshTimestamp :: Lens' VolumeRecommendation (Maybe UTCTime)

The timestamp of when the volume recommendation was last refreshed.

volumeRecommendation_currentConfiguration :: Lens' VolumeRecommendation (Maybe VolumeConfiguration)

An array of objects that describe the current configuration of the volume.

volumeRecommendation_accountId :: Lens' VolumeRecommendation (Maybe Text)

The Amazon Web Services account ID of the volume.

volumeRecommendation_utilizationMetrics :: Lens' VolumeRecommendation (Maybe [EBSUtilizationMetric])

An array of objects that describe the utilization metrics of the volume.

volumeRecommendation_volumeRecommendationOptions :: Lens' VolumeRecommendation (Maybe [VolumeRecommendationOption])

An array of objects that describe the recommendation options for the volume.

volumeRecommendation_lookBackPeriodInDays :: Lens' VolumeRecommendation (Maybe Double)

The number of days for which utilization metrics were analyzed for the volume.