Network.AWS.ComputeOptimizer.UpdateEnrollmentStatus

Description

Updates the enrollment (opt in and opt out) status of an account to the Compute Optimizer service.

If the account is a management account of an organization, this action can also be used to enroll member accounts of the organization.

You must have the appropriate permissions to opt in to Compute Optimizer, to view its recommendations, and to opt out. For more information, see Controlling access with Amazon Web Services Identity and Access Management in the Compute Optimizer User Guide.

When you opt in, Compute Optimizer automatically creates a service-linked role in your account to access its data. For more information, see Using Service-Linked Roles for Compute Optimizer in the Compute Optimizer User Guide.

Creating a Request

data UpdateEnrollmentStatus

See: newUpdateEnrollmentStatus smart constructor.

Constructors

UpdateEnrollmentStatus'

Fields

includeMemberAccounts :: Maybe Bool

Indicates whether to enroll member accounts of the organization if the account is the management account of an organization.

status :: Status

The new enrollment status of the account.

The following status options are available:

Active - Opts in your account to the Compute Optimizer service. Compute Optimizer begins analyzing the configuration and utilization metrics of your Amazon Web Services resources after you opt in. For more information, see Metrics analyzed by Compute Optimizer in the Compute Optimizer User Guide.

Inactive - Opts out your account from the Compute Optimizer service. Your account's recommendations and related metrics data will be deleted from Compute Optimizer after you opt out.

The Pending and Failed options cannot be used to update the enrollment status of an account. They are returned in the response of a request to update the enrollment status of an account.

newUpdateEnrollmentStatus

Arguments

:: Status
$sel:status:UpdateEnrollmentStatus'

-> UpdateEnrollmentStatus

Create a value of UpdateEnrollmentStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:includeMemberAccounts:UpdateEnrollmentStatus', updateEnrollmentStatus_includeMemberAccounts - Indicates whether to enroll member accounts of the organization if the account is the management account of an organization.

$sel:status:UpdateEnrollmentStatus', updateEnrollmentStatus_status - The new enrollment status of the account.

The following status options are available:

Active - Opts in your account to the Compute Optimizer service. Compute Optimizer begins analyzing the configuration and utilization metrics of your Amazon Web Services resources after you opt in. For more information, see Metrics analyzed by Compute Optimizer in the Compute Optimizer User Guide.

Inactive - Opts out your account from the Compute Optimizer service. Your account's recommendations and related metrics data will be deleted from Compute Optimizer after you opt out.

The Pending and Failed options cannot be used to update the enrollment status of an account. They are returned in the response of a request to update the enrollment status of an account.

Request Lenses

updateEnrollmentStatus_includeMemberAccounts :: Lens' UpdateEnrollmentStatus (Maybe Bool)

Indicates whether to enroll member accounts of the organization if the account is the management account of an organization.

updateEnrollmentStatus_status :: Lens' UpdateEnrollmentStatus Status

The new enrollment status of the account.

The following status options are available:

Active - Opts in your account to the Compute Optimizer service. Compute Optimizer begins analyzing the configuration and utilization metrics of your Amazon Web Services resources after you opt in. For more information, see Metrics analyzed by Compute Optimizer in the Compute Optimizer User Guide.

Inactive - Opts out your account from the Compute Optimizer service. Your account's recommendations and related metrics data will be deleted from Compute Optimizer after you opt out.

The Pending and Failed options cannot be used to update the enrollment status of an account. They are returned in the response of a request to update the enrollment status of an account.

Destructuring the Response

data UpdateEnrollmentStatusResponse

See: newUpdateEnrollmentStatusResponse smart constructor.

Constructors

UpdateEnrollmentStatusResponse'

Fields

status :: Maybe Status

The enrollment status of the account.

statusReason :: Maybe Text

The reason for the enrollment status of the account. For example, an account might show a status of Pending because member accounts of an organization require more time to be enrolled in the service.

httpStatus :: Int

The response's http status code.

newUpdateEnrollmentStatusResponse

Arguments

:: Int
$sel:httpStatus:UpdateEnrollmentStatusResponse'

-> UpdateEnrollmentStatusResponse

Create a value of UpdateEnrollmentStatusResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:UpdateEnrollmentStatus', updateEnrollmentStatusResponse_status - The enrollment status of the account.

$sel:statusReason:UpdateEnrollmentStatusResponse', updateEnrollmentStatusResponse_statusReason - The reason for the enrollment status of the account. For example, an account might show a status of Pending because member accounts of an organization require more time to be enrolled in the service.

$sel:httpStatus:UpdateEnrollmentStatusResponse', updateEnrollmentStatusResponse_httpStatus - The response's http status code.

Response Lenses

updateEnrollmentStatusResponse_status :: Lens' UpdateEnrollmentStatusResponse (Maybe Status)

The enrollment status of the account.

updateEnrollmentStatusResponse_statusReason :: Lens' UpdateEnrollmentStatusResponse (Maybe Text)

The reason for the enrollment status of the account. For example, an account might show a status of Pending because member accounts of an organization require more time to be enrolled in the service.

updateEnrollmentStatusResponse_httpStatus :: Lens' UpdateEnrollmentStatusResponse Int

The response's http status code.