Network.AWS.Config.DeleteOrganizationConfigRule

Description

Deletes the specified organization config rule and all of its evaluation results from all member accounts in that organization.

Only a master account and a delegated administrator account can delete an organization config rule. When calling this API with a delegated administrator, you must ensure Organizations ListDelegatedAdministrator permissions are added.

Config sets the state of a rule to DELETE_IN_PROGRESS until the deletion is complete. You cannot update a rule while it is in this state.

Creating a Request

data DeleteOrganizationConfigRule

See: newDeleteOrganizationConfigRule smart constructor.

Constructors

DeleteOrganizationConfigRule'

Fields

organizationConfigRuleName :: Text

The name of organization config rule that you want to delete.

newDeleteOrganizationConfigRule

Arguments

:: Text
$sel:organizationConfigRuleName:DeleteOrganizationConfigRule'

-> DeleteOrganizationConfigRule

Create a value of DeleteOrganizationConfigRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:organizationConfigRuleName:DeleteOrganizationConfigRule', deleteOrganizationConfigRule_organizationConfigRuleName - The name of organization config rule that you want to delete.

Request Lenses

deleteOrganizationConfigRule_organizationConfigRuleName :: Lens' DeleteOrganizationConfigRule Text

The name of organization config rule that you want to delete.

Destructuring the Response

data DeleteOrganizationConfigRuleResponse

See: newDeleteOrganizationConfigRuleResponse smart constructor.

Constructors

DeleteOrganizationConfigRuleResponse'

newDeleteOrganizationConfigRuleResponse :: DeleteOrganizationConfigRuleResponse

Create a value of DeleteOrganizationConfigRuleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.