Network.AWS.Config.DeleteRemediationConfiguration

Description

Deletes the remediation configuration.

Creating a Request

data DeleteRemediationConfiguration

See: newDeleteRemediationConfiguration smart constructor.

Constructors

DeleteRemediationConfiguration'

Fields

resourceType :: Maybe Text

The type of a resource.

configRuleName :: Text

The name of the Config rule for which you want to delete remediation configuration.

newDeleteRemediationConfiguration

Arguments

:: Text
$sel:configRuleName:DeleteRemediationConfiguration'

-> DeleteRemediationConfiguration

Create a value of DeleteRemediationConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceType:DeleteRemediationConfiguration', deleteRemediationConfiguration_resourceType - The type of a resource.

$sel:configRuleName:DeleteRemediationConfiguration', deleteRemediationConfiguration_configRuleName - The name of the Config rule for which you want to delete remediation configuration.

Request Lenses

deleteRemediationConfiguration_resourceType :: Lens' DeleteRemediationConfiguration (Maybe Text)

The type of a resource.

deleteRemediationConfiguration_configRuleName :: Lens' DeleteRemediationConfiguration Text

The name of the Config rule for which you want to delete remediation configuration.

Destructuring the Response

data DeleteRemediationConfigurationResponse

See: newDeleteRemediationConfigurationResponse smart constructor.

Constructors

DeleteRemediationConfigurationResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteRemediationConfigurationResponse

Arguments

:: Int
$sel:httpStatus:DeleteRemediationConfigurationResponse'

-> DeleteRemediationConfigurationResponse

Create a value of DeleteRemediationConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteRemediationConfigurationResponse', deleteRemediationConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

deleteRemediationConfigurationResponse_httpStatus :: Lens' DeleteRemediationConfigurationResponse Int

The response's http status code.