Network.AWS.Config.DescribeAggregateComplianceByConformancePacks

Description

Returns a list of the conformance packs and their associated compliance status with the count of compliant and noncompliant Config rules within each conformance pack. Also returns the total rule count which includes compliant rules, noncompliant rules, and rules that cannot be evaluated due to insufficient data.

The results can return an empty result page, but if you have a nextToken, the results are displayed on the next page.

This operation returns paginated results.

Creating a Request

data DescribeAggregateComplianceByConformancePacks

See: newDescribeAggregateComplianceByConformancePacks smart constructor.

Constructors

DescribeAggregateComplianceByConformancePacks'

Fields

filters :: Maybe AggregateConformancePackComplianceFilters

Filters the result by AggregateConformancePackComplianceFilters object.

nextToken :: Maybe Text

The nextToken string returned on a previous page that you use to get the next page of results in a paginated response.

limit :: Maybe Natural

The maximum number of conformance packs compliance details returned on each page. The default is maximum. If you specify 0, Config uses the default.

configurationAggregatorName :: Text

The name of the configuration aggregator.

newDescribeAggregateComplianceByConformancePacks

Arguments

:: Text
$sel:configurationAggregatorName:DescribeAggregateComplianceByConformancePacks'

-> DescribeAggregateComplianceByConformancePacks

Create a value of DescribeAggregateComplianceByConformancePacks with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:DescribeAggregateComplianceByConformancePacks', describeAggregateComplianceByConformancePacks_filters - Filters the result by AggregateConformancePackComplianceFilters object.

$sel:nextToken:DescribeAggregateComplianceByConformancePacks', describeAggregateComplianceByConformancePacks_nextToken - The nextToken string returned on a previous page that you use to get the next page of results in a paginated response.

$sel:limit:DescribeAggregateComplianceByConformancePacks', describeAggregateComplianceByConformancePacks_limit - The maximum number of conformance packs compliance details returned on each page. The default is maximum. If you specify 0, Config uses the default.

$sel:configurationAggregatorName:DescribeAggregateComplianceByConformancePacks', describeAggregateComplianceByConformancePacks_configurationAggregatorName - The name of the configuration aggregator.

Request Lenses

describeAggregateComplianceByConformancePacks_filters :: Lens' DescribeAggregateComplianceByConformancePacks (Maybe AggregateConformancePackComplianceFilters)

Filters the result by AggregateConformancePackComplianceFilters object.

describeAggregateComplianceByConformancePacks_nextToken :: Lens' DescribeAggregateComplianceByConformancePacks (Maybe Text)

The nextToken string returned on a previous page that you use to get the next page of results in a paginated response.

describeAggregateComplianceByConformancePacks_limit :: Lens' DescribeAggregateComplianceByConformancePacks (Maybe Natural)

The maximum number of conformance packs compliance details returned on each page. The default is maximum. If you specify 0, Config uses the default.

describeAggregateComplianceByConformancePacks_configurationAggregatorName :: Lens' DescribeAggregateComplianceByConformancePacks Text

The name of the configuration aggregator.

Destructuring the Response

data DescribeAggregateComplianceByConformancePacksResponse

See: newDescribeAggregateComplianceByConformancePacksResponse smart constructor.

Constructors

DescribeAggregateComplianceByConformancePacksResponse'

Fields

aggregateComplianceByConformancePacks :: Maybe [AggregateComplianceByConformancePack]

Returns the AggregateComplianceByConformancePack object.

nextToken :: Maybe Text

The nextToken string returned on a previous page that you use to get the next page of results in a paginated response.

httpStatus :: Int

The response's http status code.

newDescribeAggregateComplianceByConformancePacksResponse

Arguments

:: Int
$sel:httpStatus:DescribeAggregateComplianceByConformancePacksResponse'

-> DescribeAggregateComplianceByConformancePacksResponse

Create a value of DescribeAggregateComplianceByConformancePacksResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:aggregateComplianceByConformancePacks:DescribeAggregateComplianceByConformancePacksResponse', describeAggregateComplianceByConformancePacksResponse_aggregateComplianceByConformancePacks - Returns the AggregateComplianceByConformancePack object.

$sel:nextToken:DescribeAggregateComplianceByConformancePacks', describeAggregateComplianceByConformancePacksResponse_nextToken - The nextToken string returned on a previous page that you use to get the next page of results in a paginated response.

$sel:httpStatus:DescribeAggregateComplianceByConformancePacksResponse', describeAggregateComplianceByConformancePacksResponse_httpStatus - The response's http status code.

Response Lenses

describeAggregateComplianceByConformancePacksResponse_aggregateComplianceByConformancePacks :: Lens' DescribeAggregateComplianceByConformancePacksResponse (Maybe [AggregateComplianceByConformancePack])

Returns the AggregateComplianceByConformancePack object.

describeAggregateComplianceByConformancePacksResponse_nextToken :: Lens' DescribeAggregateComplianceByConformancePacksResponse (Maybe Text)

The nextToken string returned on a previous page that you use to get the next page of results in a paginated response.

describeAggregateComplianceByConformancePacksResponse_httpStatus :: Lens' DescribeAggregateComplianceByConformancePacksResponse Int

The response's http status code.