Network.AWS.Config.DescribeComplianceByConfigRule

Description

Indicates whether the specified Config rules are compliant. If a rule is noncompliant, this action returns the number of Amazon Web Services resources that do not comply with the rule.

A rule is compliant if all of the evaluated resources comply with it. It is noncompliant if any of these resources do not comply.

If Config has no current evaluation results for the rule, it returns INSUFFICIENT_DATA. This result might indicate one of the following conditions:

Config has never invoked an evaluation for the rule. To check whether it has, use the DescribeConfigRuleEvaluationStatus action to get the LastSuccessfulInvocationTime and LastFailedInvocationTime.

The rule's Lambda function is failing to send evaluation results to Config. Verify that the role you assigned to your configuration recorder includes the config:PutEvaluations permission. If the rule is a custom rule, verify that the Lambda execution role includes the config:PutEvaluations permission.

The rule's Lambda function has returned NOT_APPLICABLE for all evaluation results. This can occur if the resources were deleted or removed from the rule's scope.

This operation returns paginated results.

Creating a Request

data DescribeComplianceByConfigRule

See: newDescribeComplianceByConfigRule smart constructor.

Constructors

DescribeComplianceByConfigRule'

Fields

nextToken :: Maybe Text

The nextToken string returned on a previous page that you use to get the next page of results in a paginated response.

complianceTypes :: Maybe [ComplianceType]

Filters the results by compliance.

The allowed values are COMPLIANT and NON_COMPLIANT.

configRuleNames :: Maybe [Text]

Specify one or more Config rule names to filter the results by rule.

newDescribeComplianceByConfigRule :: DescribeComplianceByConfigRule

Create a value of DescribeComplianceByConfigRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeComplianceByConfigRule', describeComplianceByConfigRule_nextToken - The nextToken string returned on a previous page that you use to get the next page of results in a paginated response.

$sel:complianceTypes:DescribeComplianceByConfigRule', describeComplianceByConfigRule_complianceTypes - Filters the results by compliance.

The allowed values are COMPLIANT and NON_COMPLIANT.

$sel:configRuleNames:DescribeComplianceByConfigRule', describeComplianceByConfigRule_configRuleNames - Specify one or more Config rule names to filter the results by rule.

Request Lenses

describeComplianceByConfigRule_nextToken :: Lens' DescribeComplianceByConfigRule (Maybe Text)

The nextToken string returned on a previous page that you use to get the next page of results in a paginated response.

describeComplianceByConfigRule_complianceTypes :: Lens' DescribeComplianceByConfigRule (Maybe [ComplianceType])

Filters the results by compliance.

The allowed values are COMPLIANT and NON_COMPLIANT.

describeComplianceByConfigRule_configRuleNames :: Lens' DescribeComplianceByConfigRule (Maybe [Text])

Specify one or more Config rule names to filter the results by rule.

Destructuring the Response

data DescribeComplianceByConfigRuleResponse

See: newDescribeComplianceByConfigRuleResponse smart constructor.

Constructors

DescribeComplianceByConfigRuleResponse'

Fields

nextToken :: Maybe Text

The string that you use in a subsequent request to get the next page of results in a paginated response.

complianceByConfigRules :: Maybe [ComplianceByConfigRule]

Indicates whether each of the specified Config rules is compliant.

httpStatus :: Int

The response's http status code.

newDescribeComplianceByConfigRuleResponse

Arguments

:: Int
$sel:httpStatus:DescribeComplianceByConfigRuleResponse'

-> DescribeComplianceByConfigRuleResponse

Create a value of DescribeComplianceByConfigRuleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeComplianceByConfigRule', describeComplianceByConfigRuleResponse_nextToken - The string that you use in a subsequent request to get the next page of results in a paginated response.

$sel:complianceByConfigRules:DescribeComplianceByConfigRuleResponse', describeComplianceByConfigRuleResponse_complianceByConfigRules - Indicates whether each of the specified Config rules is compliant.

$sel:httpStatus:DescribeComplianceByConfigRuleResponse', describeComplianceByConfigRuleResponse_httpStatus - The response's http status code.

Response Lenses

describeComplianceByConfigRuleResponse_nextToken :: Lens' DescribeComplianceByConfigRuleResponse (Maybe Text)

The string that you use in a subsequent request to get the next page of results in a paginated response.

describeComplianceByConfigRuleResponse_complianceByConfigRules :: Lens' DescribeComplianceByConfigRuleResponse (Maybe [ComplianceByConfigRule])

Indicates whether each of the specified Config rules is compliant.

describeComplianceByConfigRuleResponse_httpStatus :: Lens' DescribeComplianceByConfigRuleResponse Int

The response's http status code.