Network.AWS.Config.DescribeConfigRules

Description

Returns details about your Config rules.

This operation returns paginated results.

Creating a Request

data DescribeConfigRules

See: newDescribeConfigRules smart constructor.

Constructors

DescribeConfigRules'

Fields

nextToken :: Maybe Text

The nextToken string returned on a previous page that you use to get the next page of results in a paginated response.

configRuleNames :: Maybe [Text]

The names of the Config rules for which you want details. If you do not specify any names, Config returns details for all your rules.

newDescribeConfigRules :: DescribeConfigRules

Create a value of DescribeConfigRules with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeConfigRules', describeConfigRules_nextToken - The nextToken string returned on a previous page that you use to get the next page of results in a paginated response.

$sel:configRuleNames:DescribeConfigRules', describeConfigRules_configRuleNames - The names of the Config rules for which you want details. If you do not specify any names, Config returns details for all your rules.

Request Lenses

describeConfigRules_nextToken :: Lens' DescribeConfigRules (Maybe Text)

The nextToken string returned on a previous page that you use to get the next page of results in a paginated response.

describeConfigRules_configRuleNames :: Lens' DescribeConfigRules (Maybe [Text])

The names of the Config rules for which you want details. If you do not specify any names, Config returns details for all your rules.

Destructuring the Response

data DescribeConfigRulesResponse

See: newDescribeConfigRulesResponse smart constructor.

Constructors

DescribeConfigRulesResponse'

Fields

nextToken :: Maybe Text

The string that you use in a subsequent request to get the next page of results in a paginated response.

configRules :: Maybe [ConfigRule]

The details about your Config rules.

httpStatus :: Int

The response's http status code.

newDescribeConfigRulesResponse

Arguments

:: Int
$sel:httpStatus:DescribeConfigRulesResponse'

-> DescribeConfigRulesResponse

Create a value of DescribeConfigRulesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeConfigRules', describeConfigRulesResponse_nextToken - The string that you use in a subsequent request to get the next page of results in a paginated response.

$sel:configRules:DescribeConfigRulesResponse', describeConfigRulesResponse_configRules - The details about your Config rules.

$sel:httpStatus:DescribeConfigRulesResponse', describeConfigRulesResponse_httpStatus - The response's http status code.

Response Lenses

describeConfigRulesResponse_nextToken :: Lens' DescribeConfigRulesResponse (Maybe Text)

The string that you use in a subsequent request to get the next page of results in a paginated response.

describeConfigRulesResponse_configRules :: Lens' DescribeConfigRulesResponse (Maybe [ConfigRule])

The details about your Config rules.

describeConfigRulesResponse_httpStatus :: Lens' DescribeConfigRulesResponse Int

The response's http status code.