Network.AWS.Config.DescribeConformancePackCompliance

Description

Returns compliance details for each rule in that conformance pack.

You must provide exact rule names.

Creating a Request

data DescribeConformancePackCompliance

See: newDescribeConformancePackCompliance smart constructor.

Constructors

DescribeConformancePackCompliance'

Fields

nextToken :: Maybe Text

The nextToken string returned in a previous request that you use to request the next page of results in a paginated response.

filters :: Maybe ConformancePackComplianceFilters

A ConformancePackComplianceFilters object.

limit :: Maybe Natural

The maximum number of Config rules within a conformance pack are returned on each page.

conformancePackName :: Text

Name of the conformance pack.

newDescribeConformancePackCompliance

Arguments

:: Text
$sel:conformancePackName:DescribeConformancePackCompliance'

-> DescribeConformancePackCompliance

Create a value of DescribeConformancePackCompliance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeConformancePackCompliance', describeConformancePackCompliance_nextToken - The nextToken string returned in a previous request that you use to request the next page of results in a paginated response.

$sel:filters:DescribeConformancePackCompliance', describeConformancePackCompliance_filters - A ConformancePackComplianceFilters object.

$sel:limit:DescribeConformancePackCompliance', describeConformancePackCompliance_limit - The maximum number of Config rules within a conformance pack are returned on each page.

$sel:conformancePackName:DescribeConformancePackCompliance', describeConformancePackCompliance_conformancePackName - Name of the conformance pack.

Request Lenses

describeConformancePackCompliance_nextToken :: Lens' DescribeConformancePackCompliance (Maybe Text)

The nextToken string returned in a previous request that you use to request the next page of results in a paginated response.

describeConformancePackCompliance_filters :: Lens' DescribeConformancePackCompliance (Maybe ConformancePackComplianceFilters)

A ConformancePackComplianceFilters object.

describeConformancePackCompliance_limit :: Lens' DescribeConformancePackCompliance (Maybe Natural)

The maximum number of Config rules within a conformance pack are returned on each page.

describeConformancePackCompliance_conformancePackName :: Lens' DescribeConformancePackCompliance Text

Name of the conformance pack.

Destructuring the Response

data DescribeConformancePackComplianceResponse

See: newDescribeConformancePackComplianceResponse smart constructor.

Constructors

DescribeConformancePackComplianceResponse'

Fields

nextToken :: Maybe Text

The nextToken string returned in a previous request that you use to request the next page of results in a paginated response.

httpStatus :: Int

The response's http status code.

conformancePackName :: Text

Name of the conformance pack.

conformancePackRuleComplianceList :: [ConformancePackRuleCompliance]

Returns a list of ConformancePackRuleCompliance objects.

newDescribeConformancePackComplianceResponse

Arguments

:: Int
$sel:httpStatus:DescribeConformancePackComplianceResponse'

-> Text
$sel:conformancePackName:DescribeConformancePackCompliance'

-> DescribeConformancePackComplianceResponse

Create a value of DescribeConformancePackComplianceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeConformancePackCompliance', describeConformancePackComplianceResponse_nextToken - The nextToken string returned in a previous request that you use to request the next page of results in a paginated response.

$sel:httpStatus:DescribeConformancePackComplianceResponse', describeConformancePackComplianceResponse_httpStatus - The response's http status code.

$sel:conformancePackName:DescribeConformancePackCompliance', describeConformancePackComplianceResponse_conformancePackName - Name of the conformance pack.

$sel:conformancePackRuleComplianceList:DescribeConformancePackComplianceResponse', describeConformancePackComplianceResponse_conformancePackRuleComplianceList - Returns a list of ConformancePackRuleCompliance objects.

Response Lenses

describeConformancePackComplianceResponse_nextToken :: Lens' DescribeConformancePackComplianceResponse (Maybe Text)

The nextToken string returned in a previous request that you use to request the next page of results in a paginated response.

describeConformancePackComplianceResponse_httpStatus :: Lens' DescribeConformancePackComplianceResponse Int

The response's http status code.

describeConformancePackComplianceResponse_conformancePackName :: Lens' DescribeConformancePackComplianceResponse Text

Name of the conformance pack.

describeConformancePackComplianceResponse_conformancePackRuleComplianceList :: Lens' DescribeConformancePackComplianceResponse [ConformancePackRuleCompliance]

Returns a list of ConformancePackRuleCompliance objects.