Network.AWS.Config.DescribeRetentionConfigurations

Description

Returns the details of one or more retention configurations. If the retention configuration name is not specified, this action returns the details for all the retention configurations for that account.

Currently, Config supports only one retention configuration per region in your account.

This operation returns paginated results.

Creating a Request

data DescribeRetentionConfigurations

See: newDescribeRetentionConfigurations smart constructor.

Constructors

DescribeRetentionConfigurations'

Fields

retentionConfigurationNames :: Maybe [Text]

A list of names of retention configurations for which you want details. If you do not specify a name, Config returns details for all the retention configurations for that account.

Currently, Config supports only one retention configuration per region in your account.

nextToken :: Maybe Text

The nextToken string returned on a previous page that you use to get the next page of results in a paginated response.

newDescribeRetentionConfigurations :: DescribeRetentionConfigurations

Create a value of DescribeRetentionConfigurations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:retentionConfigurationNames:DescribeRetentionConfigurations', describeRetentionConfigurations_retentionConfigurationNames - A list of names of retention configurations for which you want details. If you do not specify a name, Config returns details for all the retention configurations for that account.

Currently, Config supports only one retention configuration per region in your account.

$sel:nextToken:DescribeRetentionConfigurations', describeRetentionConfigurations_nextToken - The nextToken string returned on a previous page that you use to get the next page of results in a paginated response.

Request Lenses

describeRetentionConfigurations_retentionConfigurationNames :: Lens' DescribeRetentionConfigurations (Maybe [Text])

A list of names of retention configurations for which you want details. If you do not specify a name, Config returns details for all the retention configurations for that account.

Currently, Config supports only one retention configuration per region in your account.

describeRetentionConfigurations_nextToken :: Lens' DescribeRetentionConfigurations (Maybe Text)

The nextToken string returned on a previous page that you use to get the next page of results in a paginated response.

Destructuring the Response

data DescribeRetentionConfigurationsResponse

See: newDescribeRetentionConfigurationsResponse smart constructor.

Constructors

DescribeRetentionConfigurationsResponse'

Fields

retentionConfigurations :: Maybe [RetentionConfiguration]

Returns a retention configuration object.

nextToken :: Maybe Text

The nextToken string returned on a previous page that you use to get the next page of results in a paginated response.

httpStatus :: Int

The response's http status code.

newDescribeRetentionConfigurationsResponse

Arguments

:: Int
$sel:httpStatus:DescribeRetentionConfigurationsResponse'

-> DescribeRetentionConfigurationsResponse

Create a value of DescribeRetentionConfigurationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:retentionConfigurations:DescribeRetentionConfigurationsResponse', describeRetentionConfigurationsResponse_retentionConfigurations - Returns a retention configuration object.

$sel:nextToken:DescribeRetentionConfigurations', describeRetentionConfigurationsResponse_nextToken - The nextToken string returned on a previous page that you use to get the next page of results in a paginated response.

$sel:httpStatus:DescribeRetentionConfigurationsResponse', describeRetentionConfigurationsResponse_httpStatus - The response's http status code.

Response Lenses

describeRetentionConfigurationsResponse_retentionConfigurations :: Lens' DescribeRetentionConfigurationsResponse (Maybe [RetentionConfiguration])

Returns a retention configuration object.

describeRetentionConfigurationsResponse_nextToken :: Lens' DescribeRetentionConfigurationsResponse (Maybe Text)

The nextToken string returned on a previous page that you use to get the next page of results in a paginated response.

describeRetentionConfigurationsResponse_httpStatus :: Lens' DescribeRetentionConfigurationsResponse Int

The response's http status code.