Network.AWS.Config.GetAggregateConfigRuleComplianceSummary

Description

Returns the number of compliant and noncompliant rules for one or more accounts and regions in an aggregator.

The results can return an empty result page, but if you have a nextToken, the results are displayed on the next page.

Creating a Request

data GetAggregateConfigRuleComplianceSummary

See: newGetAggregateConfigRuleComplianceSummary smart constructor.

Constructors

GetAggregateConfigRuleComplianceSummary'

Fields

filters :: Maybe ConfigRuleComplianceSummaryFilters

Filters the results based on the ConfigRuleComplianceSummaryFilters object.

nextToken :: Maybe Text

The nextToken string returned on a previous page that you use to get the next page of results in a paginated response.

limit :: Maybe Natural

The maximum number of evaluation results returned on each page. The default is 1000. You cannot specify a number greater than 1000. If you specify 0, Config uses the default.

groupByKey :: Maybe ConfigRuleComplianceSummaryGroupKey

Groups the result based on ACCOUNT_ID or AWS_REGION.

configurationAggregatorName :: Text

The name of the configuration aggregator.

newGetAggregateConfigRuleComplianceSummary

Arguments

:: Text
$sel:configurationAggregatorName:GetAggregateConfigRuleComplianceSummary'

-> GetAggregateConfigRuleComplianceSummary

Create a value of GetAggregateConfigRuleComplianceSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:GetAggregateConfigRuleComplianceSummary', getAggregateConfigRuleComplianceSummary_filters - Filters the results based on the ConfigRuleComplianceSummaryFilters object.

$sel:nextToken:GetAggregateConfigRuleComplianceSummary', getAggregateConfigRuleComplianceSummary_nextToken - The nextToken string returned on a previous page that you use to get the next page of results in a paginated response.

$sel:limit:GetAggregateConfigRuleComplianceSummary', getAggregateConfigRuleComplianceSummary_limit - The maximum number of evaluation results returned on each page. The default is 1000. You cannot specify a number greater than 1000. If you specify 0, Config uses the default.

$sel:groupByKey:GetAggregateConfigRuleComplianceSummary', getAggregateConfigRuleComplianceSummary_groupByKey - Groups the result based on ACCOUNT_ID or AWS_REGION.

$sel:configurationAggregatorName:GetAggregateConfigRuleComplianceSummary', getAggregateConfigRuleComplianceSummary_configurationAggregatorName - The name of the configuration aggregator.

Request Lenses

getAggregateConfigRuleComplianceSummary_filters :: Lens' GetAggregateConfigRuleComplianceSummary (Maybe ConfigRuleComplianceSummaryFilters)

Filters the results based on the ConfigRuleComplianceSummaryFilters object.

getAggregateConfigRuleComplianceSummary_nextToken :: Lens' GetAggregateConfigRuleComplianceSummary (Maybe Text)

The nextToken string returned on a previous page that you use to get the next page of results in a paginated response.

getAggregateConfigRuleComplianceSummary_limit :: Lens' GetAggregateConfigRuleComplianceSummary (Maybe Natural)

The maximum number of evaluation results returned on each page. The default is 1000. You cannot specify a number greater than 1000. If you specify 0, Config uses the default.

getAggregateConfigRuleComplianceSummary_groupByKey :: Lens' GetAggregateConfigRuleComplianceSummary (Maybe ConfigRuleComplianceSummaryGroupKey)

Groups the result based on ACCOUNT_ID or AWS_REGION.

getAggregateConfigRuleComplianceSummary_configurationAggregatorName :: Lens' GetAggregateConfigRuleComplianceSummary Text

The name of the configuration aggregator.

Destructuring the Response

data GetAggregateConfigRuleComplianceSummaryResponse

See: newGetAggregateConfigRuleComplianceSummaryResponse smart constructor.

Constructors

GetAggregateConfigRuleComplianceSummaryResponse'

Fields

aggregateComplianceCounts :: Maybe [AggregateComplianceCount]

Returns a list of AggregateComplianceCounts object.

nextToken :: Maybe Text

The nextToken string returned on a previous page that you use to get the next page of results in a paginated response.

groupByKey :: Maybe Text

Groups the result based on ACCOUNT_ID or AWS_REGION.

httpStatus :: Int

The response's http status code.

newGetAggregateConfigRuleComplianceSummaryResponse

Arguments

:: Int
$sel:httpStatus:GetAggregateConfigRuleComplianceSummaryResponse'

-> GetAggregateConfigRuleComplianceSummaryResponse

Create a value of GetAggregateConfigRuleComplianceSummaryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:aggregateComplianceCounts:GetAggregateConfigRuleComplianceSummaryResponse', getAggregateConfigRuleComplianceSummaryResponse_aggregateComplianceCounts - Returns a list of AggregateComplianceCounts object.

$sel:nextToken:GetAggregateConfigRuleComplianceSummary', getAggregateConfigRuleComplianceSummaryResponse_nextToken - The nextToken string returned on a previous page that you use to get the next page of results in a paginated response.

$sel:groupByKey:GetAggregateConfigRuleComplianceSummary', getAggregateConfigRuleComplianceSummaryResponse_groupByKey - Groups the result based on ACCOUNT_ID or AWS_REGION.

$sel:httpStatus:GetAggregateConfigRuleComplianceSummaryResponse', getAggregateConfigRuleComplianceSummaryResponse_httpStatus - The response's http status code.

Response Lenses

getAggregateConfigRuleComplianceSummaryResponse_aggregateComplianceCounts :: Lens' GetAggregateConfigRuleComplianceSummaryResponse (Maybe [AggregateComplianceCount])

Returns a list of AggregateComplianceCounts object.

getAggregateConfigRuleComplianceSummaryResponse_nextToken :: Lens' GetAggregateConfigRuleComplianceSummaryResponse (Maybe Text)

The nextToken string returned on a previous page that you use to get the next page of results in a paginated response.

getAggregateConfigRuleComplianceSummaryResponse_groupByKey :: Lens' GetAggregateConfigRuleComplianceSummaryResponse (Maybe Text)

Groups the result based on ACCOUNT_ID or AWS_REGION.

getAggregateConfigRuleComplianceSummaryResponse_httpStatus :: Lens' GetAggregateConfigRuleComplianceSummaryResponse Int

The response's http status code.