Network.AWS.Config.GetOrganizationConfigRuleDetailedStatus

Description

Returns detailed status for each member account within an organization for a given organization config rule.

This operation returns paginated results.

Creating a Request

data GetOrganizationConfigRuleDetailedStatus

See: newGetOrganizationConfigRuleDetailedStatus smart constructor.

Constructors

GetOrganizationConfigRuleDetailedStatus'

Fields

nextToken :: Maybe Text

The nextToken string returned on a previous page that you use to get the next page of results in a paginated response.

filters :: Maybe StatusDetailFilters

A StatusDetailFilters object.

limit :: Maybe Natural

The maximum number of OrganizationConfigRuleDetailedStatus returned on each page. If you do not specify a number, Config uses the default. The default is 100.

organizationConfigRuleName :: Text

The name of organization config rule for which you want status details for member accounts.

newGetOrganizationConfigRuleDetailedStatus

Arguments

:: Text
$sel:organizationConfigRuleName:GetOrganizationConfigRuleDetailedStatus'

-> GetOrganizationConfigRuleDetailedStatus

Create a value of GetOrganizationConfigRuleDetailedStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetOrganizationConfigRuleDetailedStatus', getOrganizationConfigRuleDetailedStatus_nextToken - The nextToken string returned on a previous page that you use to get the next page of results in a paginated response.

$sel:filters:GetOrganizationConfigRuleDetailedStatus', getOrganizationConfigRuleDetailedStatus_filters - A StatusDetailFilters object.

$sel:limit:GetOrganizationConfigRuleDetailedStatus', getOrganizationConfigRuleDetailedStatus_limit - The maximum number of OrganizationConfigRuleDetailedStatus returned on each page. If you do not specify a number, Config uses the default. The default is 100.

$sel:organizationConfigRuleName:GetOrganizationConfigRuleDetailedStatus', getOrganizationConfigRuleDetailedStatus_organizationConfigRuleName - The name of organization config rule for which you want status details for member accounts.

Request Lenses

getOrganizationConfigRuleDetailedStatus_nextToken :: Lens' GetOrganizationConfigRuleDetailedStatus (Maybe Text)

The nextToken string returned on a previous page that you use to get the next page of results in a paginated response.

getOrganizationConfigRuleDetailedStatus_filters :: Lens' GetOrganizationConfigRuleDetailedStatus (Maybe StatusDetailFilters)

A StatusDetailFilters object.

getOrganizationConfigRuleDetailedStatus_limit :: Lens' GetOrganizationConfigRuleDetailedStatus (Maybe Natural)

The maximum number of OrganizationConfigRuleDetailedStatus returned on each page. If you do not specify a number, Config uses the default. The default is 100.

getOrganizationConfigRuleDetailedStatus_organizationConfigRuleName :: Lens' GetOrganizationConfigRuleDetailedStatus Text

The name of organization config rule for which you want status details for member accounts.

Destructuring the Response

data GetOrganizationConfigRuleDetailedStatusResponse

See: newGetOrganizationConfigRuleDetailedStatusResponse smart constructor.

Constructors

GetOrganizationConfigRuleDetailedStatusResponse'

Fields

nextToken :: Maybe Text

The nextToken string returned on a previous page that you use to get the next page of results in a paginated response.

organizationConfigRuleDetailedStatus :: Maybe [MemberAccountStatus]

A list of MemberAccountStatus objects.

httpStatus :: Int

The response's http status code.

newGetOrganizationConfigRuleDetailedStatusResponse

Arguments

:: Int
$sel:httpStatus:GetOrganizationConfigRuleDetailedStatusResponse'

-> GetOrganizationConfigRuleDetailedStatusResponse

Create a value of GetOrganizationConfigRuleDetailedStatusResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetOrganizationConfigRuleDetailedStatus', getOrganizationConfigRuleDetailedStatusResponse_nextToken - The nextToken string returned on a previous page that you use to get the next page of results in a paginated response.

$sel:organizationConfigRuleDetailedStatus:GetOrganizationConfigRuleDetailedStatusResponse', getOrganizationConfigRuleDetailedStatusResponse_organizationConfigRuleDetailedStatus - A list of MemberAccountStatus objects.

$sel:httpStatus:GetOrganizationConfigRuleDetailedStatusResponse', getOrganizationConfigRuleDetailedStatusResponse_httpStatus - The response's http status code.

Response Lenses

getOrganizationConfigRuleDetailedStatusResponse_nextToken :: Lens' GetOrganizationConfigRuleDetailedStatusResponse (Maybe Text)

The nextToken string returned on a previous page that you use to get the next page of results in a paginated response.

getOrganizationConfigRuleDetailedStatusResponse_organizationConfigRuleDetailedStatus :: Lens' GetOrganizationConfigRuleDetailedStatusResponse (Maybe [MemberAccountStatus])

A list of MemberAccountStatus objects.

getOrganizationConfigRuleDetailedStatusResponse_httpStatus :: Lens' GetOrganizationConfigRuleDetailedStatusResponse Int

The response's http status code.