Network.AWS.Config.GetResourceConfigHistory

Description

Returns a list of ConfigurationItems for the specified resource. The list contains details about each state of the resource during the specified time interval. If you specified a retention period to retain your ConfigurationItems between a minimum of 30 days and a maximum of 7 years (2557 days), Config returns the ConfigurationItems for the specified retention period.

The response is paginated. By default, Config returns a limit of 10 configuration items per page. You can customize this number with the limit parameter. The response includes a nextToken string. To get the next page of results, run the request again and specify the string for the nextToken parameter.

Each call to the API is limited to span a duration of seven days. It is likely that the number of records returned is smaller than the specified limit. In such cases, you can make another call, using the nextToken.

This operation returns paginated results.

Creating a Request

data GetResourceConfigHistory

The input for the GetResourceConfigHistory action.

See: newGetResourceConfigHistory smart constructor.

Constructors

GetResourceConfigHistory'

Fields

nextToken :: Maybe Text

The nextToken string returned on a previous page that you use to get the next page of results in a paginated response.

earlierTime :: Maybe POSIX

The time stamp that indicates an earlier time. If not specified, the action returns paginated results that contain configuration items that start when the first configuration item was recorded.

laterTime :: Maybe POSIX

The time stamp that indicates a later time. If not specified, current time is taken.

chronologicalOrder :: Maybe ChronologicalOrder

The chronological order for configuration items listed. By default, the results are listed in reverse chronological order.

limit :: Maybe Natural

The maximum number of configuration items returned on each page. The default is 10. You cannot specify a number greater than 100. If you specify 0, Config uses the default.

resourceType :: ResourceType

The resource type.

resourceId :: Text

The ID of the resource (for example., sg-xxxxxx).

newGetResourceConfigHistory

Arguments

:: ResourceType
$sel:resourceType:GetResourceConfigHistory'

-> Text
$sel:resourceId:GetResourceConfigHistory'

-> GetResourceConfigHistory

Create a value of GetResourceConfigHistory with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetResourceConfigHistory', getResourceConfigHistory_nextToken - The nextToken string returned on a previous page that you use to get the next page of results in a paginated response.

$sel:earlierTime:GetResourceConfigHistory', getResourceConfigHistory_earlierTime - The time stamp that indicates an earlier time. If not specified, the action returns paginated results that contain configuration items that start when the first configuration item was recorded.

$sel:laterTime:GetResourceConfigHistory', getResourceConfigHistory_laterTime - The time stamp that indicates a later time. If not specified, current time is taken.

$sel:chronologicalOrder:GetResourceConfigHistory', getResourceConfigHistory_chronologicalOrder - The chronological order for configuration items listed. By default, the results are listed in reverse chronological order.

$sel:limit:GetResourceConfigHistory', getResourceConfigHistory_limit - The maximum number of configuration items returned on each page. The default is 10. You cannot specify a number greater than 100. If you specify 0, Config uses the default.

$sel:resourceType:GetResourceConfigHistory', getResourceConfigHistory_resourceType - The resource type.

$sel:resourceId:GetResourceConfigHistory', getResourceConfigHistory_resourceId - The ID of the resource (for example., sg-xxxxxx).

Request Lenses

getResourceConfigHistory_nextToken :: Lens' GetResourceConfigHistory (Maybe Text)

The nextToken string returned on a previous page that you use to get the next page of results in a paginated response.

getResourceConfigHistory_earlierTime :: Lens' GetResourceConfigHistory (Maybe UTCTime)

The time stamp that indicates an earlier time. If not specified, the action returns paginated results that contain configuration items that start when the first configuration item was recorded.

getResourceConfigHistory_laterTime :: Lens' GetResourceConfigHistory (Maybe UTCTime)

The time stamp that indicates a later time. If not specified, current time is taken.

getResourceConfigHistory_chronologicalOrder :: Lens' GetResourceConfigHistory (Maybe ChronologicalOrder)

The chronological order for configuration items listed. By default, the results are listed in reverse chronological order.

getResourceConfigHistory_limit :: Lens' GetResourceConfigHistory (Maybe Natural)

The maximum number of configuration items returned on each page. The default is 10. You cannot specify a number greater than 100. If you specify 0, Config uses the default.

getResourceConfigHistory_resourceType :: Lens' GetResourceConfigHistory ResourceType

The resource type.

getResourceConfigHistory_resourceId :: Lens' GetResourceConfigHistory Text

The ID of the resource (for example., sg-xxxxxx).

Destructuring the Response

data GetResourceConfigHistoryResponse

The output for the GetResourceConfigHistory action.

See: newGetResourceConfigHistoryResponse smart constructor.

Constructors

GetResourceConfigHistoryResponse'

Fields

nextToken :: Maybe Text

The string that you use in a subsequent request to get the next page of results in a paginated response.

configurationItems :: Maybe [ConfigurationItem]

A list that contains the configuration history of one or more resources.

httpStatus :: Int

The response's http status code.

newGetResourceConfigHistoryResponse

Arguments

:: Int
$sel:httpStatus:GetResourceConfigHistoryResponse'

-> GetResourceConfigHistoryResponse

Create a value of GetResourceConfigHistoryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetResourceConfigHistory', getResourceConfigHistoryResponse_nextToken - The string that you use in a subsequent request to get the next page of results in a paginated response.

$sel:configurationItems:GetResourceConfigHistoryResponse', getResourceConfigHistoryResponse_configurationItems - A list that contains the configuration history of one or more resources.

$sel:httpStatus:GetResourceConfigHistoryResponse', getResourceConfigHistoryResponse_httpStatus - The response's http status code.

Response Lenses

getResourceConfigHistoryResponse_nextToken :: Lens' GetResourceConfigHistoryResponse (Maybe Text)

The string that you use in a subsequent request to get the next page of results in a paginated response.

getResourceConfigHistoryResponse_configurationItems :: Lens' GetResourceConfigHistoryResponse (Maybe [ConfigurationItem])

A list that contains the configuration history of one or more resources.

getResourceConfigHistoryResponse_httpStatus :: Lens' GetResourceConfigHistoryResponse Int

The response's http status code.