Network.AWS.Config.GetStoredQuery

Description

Returns the details of a specific stored query.

Creating a Request

data GetStoredQuery

See: newGetStoredQuery smart constructor.

Constructors

GetStoredQuery'

Fields

queryName :: Text

The name of the query.

newGetStoredQuery

Arguments

:: Text
$sel:queryName:GetStoredQuery'

-> GetStoredQuery

Create a value of GetStoredQuery with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:queryName:GetStoredQuery', getStoredQuery_queryName - The name of the query.

Request Lenses

getStoredQuery_queryName :: Lens' GetStoredQuery Text

The name of the query.

Destructuring the Response

data GetStoredQueryResponse

See: newGetStoredQueryResponse smart constructor.

Constructors

GetStoredQueryResponse'

Fields

storedQuery :: Maybe StoredQuery

Returns a StoredQuery object.

httpStatus :: Int

The response's http status code.

newGetStoredQueryResponse

Arguments

:: Int
$sel:httpStatus:GetStoredQueryResponse'

-> GetStoredQueryResponse

Create a value of GetStoredQueryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:storedQuery:GetStoredQueryResponse', getStoredQueryResponse_storedQuery - Returns a StoredQuery object.

$sel:httpStatus:GetStoredQueryResponse', getStoredQueryResponse_httpStatus - The response's http status code.

Response Lenses

getStoredQueryResponse_storedQuery :: Lens' GetStoredQueryResponse (Maybe StoredQuery)

Returns a StoredQuery object.

getStoredQueryResponse_httpStatus :: Lens' GetStoredQueryResponse Int

The response's http status code.