Network.AWS.Config.ListDiscoveredResources

Description

Accepts a resource type and returns a list of resource identifiers for the resources of that type. A resource identifier includes the resource type, ID, and (if available) the custom resource name. The results consist of resources that Config has discovered, including those that Config is not currently recording. You can narrow the results to include only resources that have specific resource IDs or a resource name.

You can specify either resource IDs or a resource name, but not both, in the same request.

The response is paginated. By default, Config lists 100 resource identifiers on each page. You can customize this number with the limit parameter. The response includes a nextToken string. To get the next page of results, run the request again and specify the string for the nextToken parameter.

This operation returns paginated results.

Creating a Request

data ListDiscoveredResources

See: newListDiscoveredResources smart constructor.

Constructors

ListDiscoveredResources'

Fields

nextToken :: Maybe Text

The nextToken string returned on a previous page that you use to get the next page of results in a paginated response.

resourceIds :: Maybe [Text]

The IDs of only those resources that you want Config to list in the response. If you do not specify this parameter, Config lists all resources of the specified type that it has discovered.

includeDeletedResources :: Maybe Bool

Specifies whether Config includes deleted resources in the results. By default, deleted resources are not included.

limit :: Maybe Natural

The maximum number of resource identifiers returned on each page. The default is 100. You cannot specify a number greater than 100. If you specify 0, Config uses the default.

resourceName :: Maybe Text

The custom name of only those resources that you want Config to list in the response. If you do not specify this parameter, Config lists all resources of the specified type that it has discovered.

resourceType :: ResourceType

The type of resources that you want Config to list in the response.

newListDiscoveredResources

Arguments

:: ResourceType
$sel:resourceType:ListDiscoveredResources'

-> ListDiscoveredResources

Create a value of ListDiscoveredResources with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDiscoveredResources', listDiscoveredResources_nextToken - The nextToken string returned on a previous page that you use to get the next page of results in a paginated response.

$sel:resourceIds:ListDiscoveredResources', listDiscoveredResources_resourceIds - The IDs of only those resources that you want Config to list in the response. If you do not specify this parameter, Config lists all resources of the specified type that it has discovered.

$sel:includeDeletedResources:ListDiscoveredResources', listDiscoveredResources_includeDeletedResources - Specifies whether Config includes deleted resources in the results. By default, deleted resources are not included.

$sel:limit:ListDiscoveredResources', listDiscoveredResources_limit - The maximum number of resource identifiers returned on each page. The default is 100. You cannot specify a number greater than 100. If you specify 0, Config uses the default.

$sel:resourceName:ListDiscoveredResources', listDiscoveredResources_resourceName - The custom name of only those resources that you want Config to list in the response. If you do not specify this parameter, Config lists all resources of the specified type that it has discovered.

$sel:resourceType:ListDiscoveredResources', listDiscoveredResources_resourceType - The type of resources that you want Config to list in the response.

Request Lenses

listDiscoveredResources_nextToken :: Lens' ListDiscoveredResources (Maybe Text)

The nextToken string returned on a previous page that you use to get the next page of results in a paginated response.

listDiscoveredResources_resourceIds :: Lens' ListDiscoveredResources (Maybe [Text])

The IDs of only those resources that you want Config to list in the response. If you do not specify this parameter, Config lists all resources of the specified type that it has discovered.

listDiscoveredResources_includeDeletedResources :: Lens' ListDiscoveredResources (Maybe Bool)

Specifies whether Config includes deleted resources in the results. By default, deleted resources are not included.

listDiscoveredResources_limit :: Lens' ListDiscoveredResources (Maybe Natural)

The maximum number of resource identifiers returned on each page. The default is 100. You cannot specify a number greater than 100. If you specify 0, Config uses the default.

listDiscoveredResources_resourceName :: Lens' ListDiscoveredResources (Maybe Text)

The custom name of only those resources that you want Config to list in the response. If you do not specify this parameter, Config lists all resources of the specified type that it has discovered.

listDiscoveredResources_resourceType :: Lens' ListDiscoveredResources ResourceType

The type of resources that you want Config to list in the response.

Destructuring the Response

data ListDiscoveredResourcesResponse

See: newListDiscoveredResourcesResponse smart constructor.

Constructors

ListDiscoveredResourcesResponse'

Fields

nextToken :: Maybe Text

The string that you use in a subsequent request to get the next page of results in a paginated response.

resourceIdentifiers :: Maybe [ResourceIdentifier]

The details that identify a resource that is discovered by Config, including the resource type, ID, and (if available) the custom resource name.

httpStatus :: Int

The response's http status code.

newListDiscoveredResourcesResponse

Arguments

:: Int
$sel:httpStatus:ListDiscoveredResourcesResponse'

-> ListDiscoveredResourcesResponse

Create a value of ListDiscoveredResourcesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDiscoveredResources', listDiscoveredResourcesResponse_nextToken - The string that you use in a subsequent request to get the next page of results in a paginated response.

$sel:resourceIdentifiers:ListDiscoveredResourcesResponse', listDiscoveredResourcesResponse_resourceIdentifiers - The details that identify a resource that is discovered by Config, including the resource type, ID, and (if available) the custom resource name.

$sel:httpStatus:ListDiscoveredResourcesResponse', listDiscoveredResourcesResponse_httpStatus - The response's http status code.

Response Lenses

listDiscoveredResourcesResponse_nextToken :: Lens' ListDiscoveredResourcesResponse (Maybe Text)

The string that you use in a subsequent request to get the next page of results in a paginated response.

listDiscoveredResourcesResponse_resourceIdentifiers :: Lens' ListDiscoveredResourcesResponse (Maybe [ResourceIdentifier])

The details that identify a resource that is discovered by Config, including the resource type, ID, and (if available) the custom resource name.

listDiscoveredResourcesResponse_httpStatus :: Lens' ListDiscoveredResourcesResponse Int

The response's http status code.