Network.AWS.Config.PutConfigurationAggregator

Description

Creates and updates the configuration aggregator with the selected source accounts and regions. The source account can be individual account(s) or an organization.

accountIds that are passed will be replaced with existing accounts. If you want to add additional accounts into the aggregator, call DescribeAggregator to get the previous accounts and then append new ones.

Config should be enabled in source accounts and regions you want to aggregate.

If your source type is an organization, you must be signed in to the management account or a registered delegated administrator and all the features must be enabled in your organization. If the caller is a management account, Config calls EnableAwsServiceAccess API to enable integration between Config and Organizations. If the caller is a registered delegated administrator, Config calls ListDelegatedAdministrators API to verify whether the caller is a valid delegated administrator.

To register a delegated administrator, see Register a Delegated Administrator in the Config developer guide.

Creating a Request

data PutConfigurationAggregator

See: newPutConfigurationAggregator smart constructor.

Constructors

PutConfigurationAggregator'

Fields

tags :: Maybe [Tag]

An array of tag object.

accountAggregationSources :: Maybe [AccountAggregationSource]

A list of AccountAggregationSource object.

organizationAggregationSource :: Maybe OrganizationAggregationSource

An OrganizationAggregationSource object.

configurationAggregatorName :: Text

The name of the configuration aggregator.

newPutConfigurationAggregator

Arguments

:: Text
$sel:configurationAggregatorName:PutConfigurationAggregator'

-> PutConfigurationAggregator

Create a value of PutConfigurationAggregator with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:PutConfigurationAggregator', putConfigurationAggregator_tags - An array of tag object.

$sel:accountAggregationSources:PutConfigurationAggregator', putConfigurationAggregator_accountAggregationSources - A list of AccountAggregationSource object.

$sel:organizationAggregationSource:PutConfigurationAggregator', putConfigurationAggregator_organizationAggregationSource - An OrganizationAggregationSource object.

$sel:configurationAggregatorName:PutConfigurationAggregator', putConfigurationAggregator_configurationAggregatorName - The name of the configuration aggregator.

Request Lenses

putConfigurationAggregator_tags :: Lens' PutConfigurationAggregator (Maybe [Tag])

An array of tag object.

putConfigurationAggregator_accountAggregationSources :: Lens' PutConfigurationAggregator (Maybe [AccountAggregationSource])

A list of AccountAggregationSource object.

putConfigurationAggregator_organizationAggregationSource :: Lens' PutConfigurationAggregator (Maybe OrganizationAggregationSource)

An OrganizationAggregationSource object.

putConfigurationAggregator_configurationAggregatorName :: Lens' PutConfigurationAggregator Text

The name of the configuration aggregator.

Destructuring the Response

data PutConfigurationAggregatorResponse

See: newPutConfigurationAggregatorResponse smart constructor.

Constructors

PutConfigurationAggregatorResponse'

Fields

configurationAggregator :: Maybe ConfigurationAggregator

Returns a ConfigurationAggregator object.

httpStatus :: Int

The response's http status code.

newPutConfigurationAggregatorResponse

Arguments

:: Int
$sel:httpStatus:PutConfigurationAggregatorResponse'

-> PutConfigurationAggregatorResponse

Create a value of PutConfigurationAggregatorResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configurationAggregator:PutConfigurationAggregatorResponse', putConfigurationAggregatorResponse_configurationAggregator - Returns a ConfigurationAggregator object.

$sel:httpStatus:PutConfigurationAggregatorResponse', putConfigurationAggregatorResponse_httpStatus - The response's http status code.

Response Lenses

putConfigurationAggregatorResponse_configurationAggregator :: Lens' PutConfigurationAggregatorResponse (Maybe ConfigurationAggregator)

Returns a ConfigurationAggregator object.

putConfigurationAggregatorResponse_httpStatus :: Lens' PutConfigurationAggregatorResponse Int

The response's http status code.