Network.AWS.Config.PutConformancePack

Description

Creates or updates a conformance pack. A conformance pack is a collection of Config rules that can be easily deployed in an account and a region and across Amazon Web Services Organization.

This API creates a service linked role AWSServiceRoleForConfigConforms in your account. The service linked role is created only when the role does not exist in your account.

You must specify either the TemplateS3Uri or the TemplateBody parameter, but not both. If you provide both Config uses the TemplateS3Uri parameter and ignores the TemplateBody parameter.

Creating a Request

data PutConformancePack

See: newPutConformancePack smart constructor.

Constructors

PutConformancePack'

Fields

templateS3Uri :: Maybe Text

Location of file containing the template body (s3://bucketname/prefix). The uri must point to the conformance pack template (max size: 300 KB) that is located in an Amazon S3 bucket in the same region as the conformance pack.

You must have access to read Amazon S3 bucket.

deliveryS3Bucket :: Maybe Text

The name of the Amazon S3 bucket where Config stores conformance pack templates.

This field is optional.

deliveryS3KeyPrefix :: Maybe Text

The prefix for the Amazon S3 bucket.

This field is optional.

templateBody :: Maybe Text

A string containing full conformance pack template body. Structure containing the template body with a minimum length of 1 byte and a maximum length of 51,200 bytes.

You can only use a YAML template with one resource type, that is, config rule and a remediation action.

conformancePackInputParameters :: Maybe [ConformancePackInputParameter]

A list of ConformancePackInputParameter objects.

conformancePackName :: Text

Name of the conformance pack you want to create.

newPutConformancePack

Arguments

:: Text
$sel:conformancePackName:PutConformancePack'

-> PutConformancePack

Create a value of PutConformancePack with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:templateS3Uri:PutConformancePack', putConformancePack_templateS3Uri - Location of file containing the template body (s3://bucketname/prefix). The uri must point to the conformance pack template (max size: 300 KB) that is located in an Amazon S3 bucket in the same region as the conformance pack.

You must have access to read Amazon S3 bucket.

$sel:deliveryS3Bucket:PutConformancePack', putConformancePack_deliveryS3Bucket - The name of the Amazon S3 bucket where Config stores conformance pack templates.

This field is optional.

$sel:deliveryS3KeyPrefix:PutConformancePack', putConformancePack_deliveryS3KeyPrefix - The prefix for the Amazon S3 bucket.

This field is optional.

$sel:templateBody:PutConformancePack', putConformancePack_templateBody - A string containing full conformance pack template body. Structure containing the template body with a minimum length of 1 byte and a maximum length of 51,200 bytes.

You can only use a YAML template with one resource type, that is, config rule and a remediation action.

$sel:conformancePackInputParameters:PutConformancePack', putConformancePack_conformancePackInputParameters - A list of ConformancePackInputParameter objects.

$sel:conformancePackName:PutConformancePack', putConformancePack_conformancePackName - Name of the conformance pack you want to create.

Request Lenses

putConformancePack_templateS3Uri :: Lens' PutConformancePack (Maybe Text)

Location of file containing the template body (s3://bucketname/prefix). The uri must point to the conformance pack template (max size: 300 KB) that is located in an Amazon S3 bucket in the same region as the conformance pack.

You must have access to read Amazon S3 bucket.

putConformancePack_deliveryS3Bucket :: Lens' PutConformancePack (Maybe Text)

The name of the Amazon S3 bucket where Config stores conformance pack templates.

This field is optional.

putConformancePack_deliveryS3KeyPrefix :: Lens' PutConformancePack (Maybe Text)

The prefix for the Amazon S3 bucket.

This field is optional.

putConformancePack_templateBody :: Lens' PutConformancePack (Maybe Text)

A string containing full conformance pack template body. Structure containing the template body with a minimum length of 1 byte and a maximum length of 51,200 bytes.

You can only use a YAML template with one resource type, that is, config rule and a remediation action.

putConformancePack_conformancePackInputParameters :: Lens' PutConformancePack (Maybe [ConformancePackInputParameter])

A list of ConformancePackInputParameter objects.

putConformancePack_conformancePackName :: Lens' PutConformancePack Text

Name of the conformance pack you want to create.

Destructuring the Response

data PutConformancePackResponse

See: newPutConformancePackResponse smart constructor.

Constructors

PutConformancePackResponse'

Fields

conformancePackArn :: Maybe Text

ARN of the conformance pack.

httpStatus :: Int

The response's http status code.

newPutConformancePackResponse

Arguments

:: Int
$sel:httpStatus:PutConformancePackResponse'

-> PutConformancePackResponse

Create a value of PutConformancePackResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:conformancePackArn:PutConformancePackResponse', putConformancePackResponse_conformancePackArn - ARN of the conformance pack.

$sel:httpStatus:PutConformancePackResponse', putConformancePackResponse_httpStatus - The response's http status code.

Response Lenses

putConformancePackResponse_conformancePackArn :: Lens' PutConformancePackResponse (Maybe Text)

ARN of the conformance pack.

putConformancePackResponse_httpStatus :: Lens' PutConformancePackResponse Int

The response's http status code.