Network.AWS.Config.PutEvaluations

Description

Used by an Lambda function to deliver evaluation results to Config. This action is required in every Lambda function that is invoked by an Config rule.

Creating a Request

data PutEvaluations

See: newPutEvaluations smart constructor.

Constructors

PutEvaluations'

Fields

evaluations :: Maybe [Evaluation]

The assessments that the Lambda function performs. Each evaluation identifies an Amazon Web Services resource and indicates whether it complies with the Config rule that invokes the Lambda function.

testMode :: Maybe Bool

Use this parameter to specify a test run for PutEvaluations. You can verify whether your Lambda function will deliver evaluation results to Config. No updates occur to your existing evaluations, and evaluation results are not sent to Config.

When TestMode is true, PutEvaluations doesn't require a valid value for the ResultToken parameter, but the value cannot be null.

resultToken :: Text

An encrypted token that associates an evaluation with an Config rule. Identifies the rule and the event that triggered the evaluation.

newPutEvaluations

Arguments

:: Text
$sel:resultToken:PutEvaluations'

-> PutEvaluations

Create a value of PutEvaluations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:evaluations:PutEvaluations', putEvaluations_evaluations - The assessments that the Lambda function performs. Each evaluation identifies an Amazon Web Services resource and indicates whether it complies with the Config rule that invokes the Lambda function.

$sel:testMode:PutEvaluations', putEvaluations_testMode - Use this parameter to specify a test run for PutEvaluations. You can verify whether your Lambda function will deliver evaluation results to Config. No updates occur to your existing evaluations, and evaluation results are not sent to Config.

When TestMode is true, PutEvaluations doesn't require a valid value for the ResultToken parameter, but the value cannot be null.

$sel:resultToken:PutEvaluations', putEvaluations_resultToken - An encrypted token that associates an evaluation with an Config rule. Identifies the rule and the event that triggered the evaluation.

Request Lenses

putEvaluations_evaluations :: Lens' PutEvaluations (Maybe [Evaluation])

The assessments that the Lambda function performs. Each evaluation identifies an Amazon Web Services resource and indicates whether it complies with the Config rule that invokes the Lambda function.

putEvaluations_testMode :: Lens' PutEvaluations (Maybe Bool)

Use this parameter to specify a test run for PutEvaluations. You can verify whether your Lambda function will deliver evaluation results to Config. No updates occur to your existing evaluations, and evaluation results are not sent to Config.

When TestMode is true, PutEvaluations doesn't require a valid value for the ResultToken parameter, but the value cannot be null.

putEvaluations_resultToken :: Lens' PutEvaluations Text

An encrypted token that associates an evaluation with an Config rule. Identifies the rule and the event that triggered the evaluation.

Destructuring the Response

data PutEvaluationsResponse

See: newPutEvaluationsResponse smart constructor.

Constructors

PutEvaluationsResponse'

Fields

failedEvaluations :: Maybe [Evaluation]

Requests that failed because of a client or server error.

httpStatus :: Int

The response's http status code.

newPutEvaluationsResponse

Arguments

:: Int
$sel:httpStatus:PutEvaluationsResponse'

-> PutEvaluationsResponse

Create a value of PutEvaluationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failedEvaluations:PutEvaluationsResponse', putEvaluationsResponse_failedEvaluations - Requests that failed because of a client or server error.

$sel:httpStatus:PutEvaluationsResponse', putEvaluationsResponse_httpStatus - The response's http status code.

Response Lenses

putEvaluationsResponse_failedEvaluations :: Lens' PutEvaluationsResponse (Maybe [Evaluation])

Requests that failed because of a client or server error.

putEvaluationsResponse_httpStatus :: Lens' PutEvaluationsResponse Int

The response's http status code.