Network.AWS.Config.PutOrganizationConformancePack

Description

Deploys conformance packs across member accounts in an Amazon Web Services Organization.

Only a master account and a delegated administrator can call this API. When calling this API with a delegated administrator, you must ensure Organizations ListDelegatedAdministrator permissions are added.

This API enables organization service access for config-multiaccountsetup.amazonaws.com through the EnableAWSServiceAccess action and creates a service linked role AWSServiceRoleForConfigMultiAccountSetup in the master or delegated administrator account of your organization. The service linked role is created only when the role does not exist in the caller account. To use this API with delegated administrator, register a delegated administrator by calling Amazon Web Services Organization register-delegate-admin for config-multiaccountsetup.amazonaws.com.

Prerequisite: Ensure you call EnableAllFeatures API to enable all features in an organization.

You must specify either the TemplateS3Uri or the TemplateBody parameter, but not both. If you provide both Config uses the TemplateS3Uri parameter and ignores the TemplateBody parameter.

Config sets the state of a conformance pack to CREATE_IN_PROGRESS and UPDATE_IN_PROGRESS until the conformance pack is created or updated. You cannot update a conformance pack while it is in this state.

You can create 50 conformance packs with 25 Config rules in each pack and 3 delegated administrator per organization.

Creating a Request

data PutOrganizationConformancePack

See: newPutOrganizationConformancePack smart constructor.

Constructors

PutOrganizationConformancePack'

Fields

templateS3Uri :: Maybe Text

Location of file containing the template body. The uri must point to the conformance pack template (max size: 300 KB).

You must have access to read Amazon S3 bucket.

deliveryS3Bucket :: Maybe Text

The name of the Amazon S3 bucket where Config stores conformance pack templates.

This field is optional. If used, it must be prefixed with awsconfigconforms.

deliveryS3KeyPrefix :: Maybe Text

The prefix for the Amazon S3 bucket.

This field is optional.

templateBody :: Maybe Text

A string containing full conformance pack template body. Structure containing the template body with a minimum length of 1 byte and a maximum length of 51,200 bytes.

conformancePackInputParameters :: Maybe [ConformancePackInputParameter]

A list of ConformancePackInputParameter objects.

excludedAccounts :: Maybe [Text]

A list of Amazon Web Services accounts to be excluded from an organization conformance pack while deploying a conformance pack.

organizationConformancePackName :: Text

Name of the organization conformance pack you want to create.

newPutOrganizationConformancePack

Arguments

:: Text
$sel:organizationConformancePackName:PutOrganizationConformancePack'

-> PutOrganizationConformancePack

Create a value of PutOrganizationConformancePack with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:templateS3Uri:PutOrganizationConformancePack', putOrganizationConformancePack_templateS3Uri - Location of file containing the template body. The uri must point to the conformance pack template (max size: 300 KB).

You must have access to read Amazon S3 bucket.

$sel:deliveryS3Bucket:PutOrganizationConformancePack', putOrganizationConformancePack_deliveryS3Bucket - The name of the Amazon S3 bucket where Config stores conformance pack templates.

This field is optional. If used, it must be prefixed with awsconfigconforms.

$sel:deliveryS3KeyPrefix:PutOrganizationConformancePack', putOrganizationConformancePack_deliveryS3KeyPrefix - The prefix for the Amazon S3 bucket.

This field is optional.

$sel:templateBody:PutOrganizationConformancePack', putOrganizationConformancePack_templateBody - A string containing full conformance pack template body. Structure containing the template body with a minimum length of 1 byte and a maximum length of 51,200 bytes.

$sel:conformancePackInputParameters:PutOrganizationConformancePack', putOrganizationConformancePack_conformancePackInputParameters - A list of ConformancePackInputParameter objects.

$sel:excludedAccounts:PutOrganizationConformancePack', putOrganizationConformancePack_excludedAccounts - A list of Amazon Web Services accounts to be excluded from an organization conformance pack while deploying a conformance pack.

$sel:organizationConformancePackName:PutOrganizationConformancePack', putOrganizationConformancePack_organizationConformancePackName - Name of the organization conformance pack you want to create.

Request Lenses

putOrganizationConformancePack_templateS3Uri :: Lens' PutOrganizationConformancePack (Maybe Text)

Location of file containing the template body. The uri must point to the conformance pack template (max size: 300 KB).

You must have access to read Amazon S3 bucket.

putOrganizationConformancePack_deliveryS3Bucket :: Lens' PutOrganizationConformancePack (Maybe Text)

The name of the Amazon S3 bucket where Config stores conformance pack templates.

This field is optional. If used, it must be prefixed with awsconfigconforms.

putOrganizationConformancePack_deliveryS3KeyPrefix :: Lens' PutOrganizationConformancePack (Maybe Text)

The prefix for the Amazon S3 bucket.

This field is optional.

putOrganizationConformancePack_templateBody :: Lens' PutOrganizationConformancePack (Maybe Text)

A string containing full conformance pack template body. Structure containing the template body with a minimum length of 1 byte and a maximum length of 51,200 bytes.

putOrganizationConformancePack_conformancePackInputParameters :: Lens' PutOrganizationConformancePack (Maybe [ConformancePackInputParameter])

A list of ConformancePackInputParameter objects.

putOrganizationConformancePack_excludedAccounts :: Lens' PutOrganizationConformancePack (Maybe [Text])

A list of Amazon Web Services accounts to be excluded from an organization conformance pack while deploying a conformance pack.

putOrganizationConformancePack_organizationConformancePackName :: Lens' PutOrganizationConformancePack Text

Name of the organization conformance pack you want to create.

Destructuring the Response

data PutOrganizationConformancePackResponse

See: newPutOrganizationConformancePackResponse smart constructor.

Constructors

PutOrganizationConformancePackResponse'

Fields

organizationConformancePackArn :: Maybe Text

ARN of the organization conformance pack.

httpStatus :: Int

The response's http status code.

newPutOrganizationConformancePackResponse

Arguments

:: Int
$sel:httpStatus:PutOrganizationConformancePackResponse'

-> PutOrganizationConformancePackResponse

Create a value of PutOrganizationConformancePackResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:organizationConformancePackArn:PutOrganizationConformancePackResponse', putOrganizationConformancePackResponse_organizationConformancePackArn - ARN of the organization conformance pack.

$sel:httpStatus:PutOrganizationConformancePackResponse', putOrganizationConformancePackResponse_httpStatus - The response's http status code.

Response Lenses

putOrganizationConformancePackResponse_organizationConformancePackArn :: Lens' PutOrganizationConformancePackResponse (Maybe Text)

ARN of the organization conformance pack.

putOrganizationConformancePackResponse_httpStatus :: Lens' PutOrganizationConformancePackResponse Int

The response's http status code.