Network.AWS.Config.PutRetentionConfiguration

Description

Creates and updates the retention configuration with details about retention period (number of days) that Config stores your historical information. The API creates the RetentionConfiguration object and names the object as default. When you have a RetentionConfiguration object named default, calling the API modifies the default object.

Currently, Config supports only one retention configuration per region in your account.

Creating a Request

data PutRetentionConfiguration

See: newPutRetentionConfiguration smart constructor.

Constructors

PutRetentionConfiguration'

Fields

retentionPeriodInDays :: Natural

Number of days Config stores your historical information.

Currently, only applicable to the configuration item history.

newPutRetentionConfiguration

Arguments

:: Natural
$sel:retentionPeriodInDays:PutRetentionConfiguration'

-> PutRetentionConfiguration

Create a value of PutRetentionConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:retentionPeriodInDays:PutRetentionConfiguration', putRetentionConfiguration_retentionPeriodInDays - Number of days Config stores your historical information.

Currently, only applicable to the configuration item history.

Request Lenses

putRetentionConfiguration_retentionPeriodInDays :: Lens' PutRetentionConfiguration Natural

Number of days Config stores your historical information.

Currently, only applicable to the configuration item history.

Destructuring the Response

data PutRetentionConfigurationResponse

See: newPutRetentionConfigurationResponse smart constructor.

Constructors

PutRetentionConfigurationResponse'

Fields

retentionConfiguration :: Maybe RetentionConfiguration

Returns a retention configuration object.

httpStatus :: Int

The response's http status code.

newPutRetentionConfigurationResponse

Arguments

:: Int
$sel:httpStatus:PutRetentionConfigurationResponse'

-> PutRetentionConfigurationResponse

Create a value of PutRetentionConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:retentionConfiguration:PutRetentionConfigurationResponse', putRetentionConfigurationResponse_retentionConfiguration - Returns a retention configuration object.

$sel:httpStatus:PutRetentionConfigurationResponse', putRetentionConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

putRetentionConfigurationResponse_retentionConfiguration :: Lens' PutRetentionConfigurationResponse (Maybe RetentionConfiguration)

Returns a retention configuration object.

putRetentionConfigurationResponse_httpStatus :: Lens' PutRetentionConfigurationResponse Int

The response's http status code.