Network.AWS.Config.SelectResourceConfig

Description

Accepts a structured query language (SQL) SELECT command, performs the corresponding search, and returns resource configurations matching the properties.

For more information about query components, see the Query Components section in the Config Developer Guide.

This operation returns paginated results.

Creating a Request

data SelectResourceConfig

See: newSelectResourceConfig smart constructor.

Constructors

SelectResourceConfig'

Fields

nextToken :: Maybe Text

The nextToken string returned in a previous request that you use to request the next page of results in a paginated response.

limit :: Maybe Natural

The maximum number of query results returned on each page.

expression :: Text

The SQL query SELECT command.

newSelectResourceConfig

Arguments

:: Text
$sel:expression:SelectResourceConfig'

-> SelectResourceConfig

Create a value of SelectResourceConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:SelectResourceConfig', selectResourceConfig_nextToken - The nextToken string returned in a previous request that you use to request the next page of results in a paginated response.

$sel:limit:SelectResourceConfig', selectResourceConfig_limit - The maximum number of query results returned on each page.

$sel:expression:SelectResourceConfig', selectResourceConfig_expression - The SQL query SELECT command.

Request Lenses

selectResourceConfig_nextToken :: Lens' SelectResourceConfig (Maybe Text)

The nextToken string returned in a previous request that you use to request the next page of results in a paginated response.

selectResourceConfig_limit :: Lens' SelectResourceConfig (Maybe Natural)

The maximum number of query results returned on each page.

selectResourceConfig_expression :: Lens' SelectResourceConfig Text

The SQL query SELECT command.

Destructuring the Response

data SelectResourceConfigResponse

See: newSelectResourceConfigResponse smart constructor.

Constructors

SelectResourceConfigResponse'

Fields

results :: Maybe [Text]

Returns the results for the SQL query.

queryInfo :: Maybe QueryInfo

Returns the QueryInfo object.

nextToken :: Maybe Text

The nextToken string returned in a previous request that you use to request the next page of results in a paginated response.

httpStatus :: Int

The response's http status code.

newSelectResourceConfigResponse

Arguments

:: Int
$sel:httpStatus:SelectResourceConfigResponse'

-> SelectResourceConfigResponse

Create a value of SelectResourceConfigResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:results:SelectResourceConfigResponse', selectResourceConfigResponse_results - Returns the results for the SQL query.

$sel:queryInfo:SelectResourceConfigResponse', selectResourceConfigResponse_queryInfo - Returns the QueryInfo object.

$sel:nextToken:SelectResourceConfig', selectResourceConfigResponse_nextToken - The nextToken string returned in a previous request that you use to request the next page of results in a paginated response.

$sel:httpStatus:SelectResourceConfigResponse', selectResourceConfigResponse_httpStatus - The response's http status code.

Response Lenses

selectResourceConfigResponse_results :: Lens' SelectResourceConfigResponse (Maybe [Text])

Returns the results for the SQL query.

selectResourceConfigResponse_queryInfo :: Lens' SelectResourceConfigResponse (Maybe QueryInfo)

Returns the QueryInfo object.

selectResourceConfigResponse_nextToken :: Lens' SelectResourceConfigResponse (Maybe Text)

The nextToken string returned in a previous request that you use to request the next page of results in a paginated response.

selectResourceConfigResponse_httpStatus :: Lens' SelectResourceConfigResponse Int

The response's http status code.