Network.AWS.Config.Types.AccountAggregationSource

Description

Documentation

data AccountAggregationSource

A collection of accounts and regions.

See: newAccountAggregationSource smart constructor.

Constructors

AccountAggregationSource'

Fields

allAwsRegions :: Maybe Bool

If true, aggregate existing Config regions and future regions.

awsRegions :: Maybe (NonEmpty Text)

The source regions being aggregated.

accountIds :: NonEmpty Text

The 12-digit account ID of the account being aggregated.

newAccountAggregationSource

Arguments

:: NonEmpty Text
$sel:accountIds:AccountAggregationSource'

-> AccountAggregationSource

Create a value of AccountAggregationSource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:allAwsRegions:AccountAggregationSource', accountAggregationSource_allAwsRegions - If true, aggregate existing Config regions and future regions.

$sel:awsRegions:AccountAggregationSource', accountAggregationSource_awsRegions - The source regions being aggregated.

$sel:accountIds:AccountAggregationSource', accountAggregationSource_accountIds - The 12-digit account ID of the account being aggregated.

accountAggregationSource_allAwsRegions :: Lens' AccountAggregationSource (Maybe Bool)

If true, aggregate existing Config regions and future regions.

accountAggregationSource_awsRegions :: Lens' AccountAggregationSource (Maybe (NonEmpty Text))

The source regions being aggregated.

accountAggregationSource_accountIds :: Lens' AccountAggregationSource (NonEmpty Text)

The 12-digit account ID of the account being aggregated.