Network.AWS.Config.Types.AggregateConformancePackComplianceSummary

Description

Documentation

data AggregateConformancePackComplianceSummary

Provides a summary of compliance based on either account ID or region.

See: newAggregateConformancePackComplianceSummary smart constructor.

Constructors

AggregateConformancePackComplianceSummary'

Fields

groupName :: Maybe Text

Groups the result based on Amazon Web Services account ID or Amazon Web Services Region.

complianceSummary :: Maybe AggregateConformancePackComplianceCount

Returns an AggregateConformancePackComplianceCount object.

newAggregateConformancePackComplianceSummary :: AggregateConformancePackComplianceSummary

Create a value of AggregateConformancePackComplianceSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groupName:AggregateConformancePackComplianceSummary', aggregateConformancePackComplianceSummary_groupName - Groups the result based on Amazon Web Services account ID or Amazon Web Services Region.

$sel:complianceSummary:AggregateConformancePackComplianceSummary', aggregateConformancePackComplianceSummary_complianceSummary - Returns an AggregateConformancePackComplianceCount object.

aggregateConformancePackComplianceSummary_groupName :: Lens' AggregateConformancePackComplianceSummary (Maybe Text)

Groups the result based on Amazon Web Services account ID or Amazon Web Services Region.

aggregateConformancePackComplianceSummary_complianceSummary :: Lens' AggregateConformancePackComplianceSummary (Maybe AggregateConformancePackComplianceCount)

Returns an AggregateConformancePackComplianceCount object.