Network.AWS.Config.Types.AggregatedSourceStatus

Description

Documentation

data AggregatedSourceStatus

The current sync status between the source and the aggregator account.

See: newAggregatedSourceStatus smart constructor.

Constructors

AggregatedSourceStatus'

Fields

lastErrorCode :: Maybe Text

The error code that Config returned when the source account aggregation last failed.

lastUpdateStatus :: Maybe AggregatedSourceStatusType

Filters the last updated status type.

Valid value FAILED indicates errors while moving data.

Valid value SUCCEEDED indicates the data was successfully moved.

Valid value OUTDATED indicates the data is not the most recent.

sourceType :: Maybe AggregatedSourceType

The source account or an organization.

sourceId :: Maybe Text

The source account ID or an organization.

lastErrorMessage :: Maybe Text

The message indicating that the source account aggregation failed due to an error.

awsRegion :: Maybe Text

The region authorized to collect aggregated data.

lastUpdateTime :: Maybe POSIX

The time of the last update.

newAggregatedSourceStatus :: AggregatedSourceStatus

Create a value of AggregatedSourceStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastErrorCode:AggregatedSourceStatus', aggregatedSourceStatus_lastErrorCode - The error code that Config returned when the source account aggregation last failed.

$sel:lastUpdateStatus:AggregatedSourceStatus', aggregatedSourceStatus_lastUpdateStatus - Filters the last updated status type.

Valid value FAILED indicates errors while moving data.

Valid value SUCCEEDED indicates the data was successfully moved.

Valid value OUTDATED indicates the data is not the most recent.

$sel:sourceType:AggregatedSourceStatus', aggregatedSourceStatus_sourceType - The source account or an organization.

$sel:sourceId:AggregatedSourceStatus', aggregatedSourceStatus_sourceId - The source account ID or an organization.

$sel:lastErrorMessage:AggregatedSourceStatus', aggregatedSourceStatus_lastErrorMessage - The message indicating that the source account aggregation failed due to an error.

$sel:awsRegion:AggregatedSourceStatus', aggregatedSourceStatus_awsRegion - The region authorized to collect aggregated data.

$sel:lastUpdateTime:AggregatedSourceStatus', aggregatedSourceStatus_lastUpdateTime - The time of the last update.

aggregatedSourceStatus_lastErrorCode :: Lens' AggregatedSourceStatus (Maybe Text)

The error code that Config returned when the source account aggregation last failed.

aggregatedSourceStatus_lastUpdateStatus :: Lens' AggregatedSourceStatus (Maybe AggregatedSourceStatusType)

Filters the last updated status type.

Valid value FAILED indicates errors while moving data.

Valid value SUCCEEDED indicates the data was successfully moved.

Valid value OUTDATED indicates the data is not the most recent.

aggregatedSourceStatus_sourceType :: Lens' AggregatedSourceStatus (Maybe AggregatedSourceType)

The source account or an organization.

aggregatedSourceStatus_sourceId :: Lens' AggregatedSourceStatus (Maybe Text)

The source account ID or an organization.

aggregatedSourceStatus_lastErrorMessage :: Lens' AggregatedSourceStatus (Maybe Text)

The message indicating that the source account aggregation failed due to an error.

aggregatedSourceStatus_awsRegion :: Lens' AggregatedSourceStatus (Maybe Text)

The region authorized to collect aggregated data.

aggregatedSourceStatus_lastUpdateTime :: Lens' AggregatedSourceStatus (Maybe UTCTime)

The time of the last update.