Network.AWS.Config.Types.BaseConfigurationItem

Description

Documentation

data BaseConfigurationItem

The detailed configuration of a specified resource.

See: newBaseConfigurationItem smart constructor.

Constructors

BaseConfigurationItem'

Fields

resourceId :: Maybe Text

The ID of the resource (for example., sg-xxxxxx).

accountId :: Maybe Text

The 12-digit Amazon Web Services account ID associated with the resource.

configuration :: Maybe Text

The description of the resource configuration.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the resource.

configurationStateId :: Maybe Text

An identifier that indicates the ordering of the configuration items of a resource.

version :: Maybe Text

The version number of the resource configuration.

resourceType :: Maybe ResourceType

The type of Amazon Web Services resource.

supplementaryConfiguration :: Maybe (HashMap Text Text)

Configuration attributes that Config returns for certain resource types to supplement the information returned for the configuration parameter.

availabilityZone :: Maybe Text

The Availability Zone associated with the resource.

configurationItemCaptureTime :: Maybe POSIX

The time when the configuration recording was initiated.

configurationItemStatus :: Maybe ConfigurationItemStatus

The configuration item status. The valid values are:

OK – The resource configuration has been updated

ResourceDiscovered – The resource was newly discovered

ResourceNotRecorded – The resource was discovered but its configuration was not recorded since the recorder excludes the recording of resources of this type

ResourceDeleted – The resource was deleted

ResourceDeletedNotRecorded – The resource was deleted but its configuration was not recorded since the recorder excludes the recording of resources of this type

The CIs do not incur any cost.

resourceCreationTime :: Maybe POSIX

The time stamp when the resource was created.

resourceName :: Maybe Text

The custom name of the resource, if available.

awsRegion :: Maybe Text

The region where the resource resides.

newBaseConfigurationItem :: BaseConfigurationItem

Create a value of BaseConfigurationItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceId:BaseConfigurationItem', baseConfigurationItem_resourceId - The ID of the resource (for example., sg-xxxxxx).

$sel:accountId:BaseConfigurationItem', baseConfigurationItem_accountId - The 12-digit Amazon Web Services account ID associated with the resource.

$sel:configuration:BaseConfigurationItem', baseConfigurationItem_configuration - The description of the resource configuration.

$sel:arn:BaseConfigurationItem', baseConfigurationItem_arn - The Amazon Resource Name (ARN) of the resource.

$sel:configurationStateId:BaseConfigurationItem', baseConfigurationItem_configurationStateId - An identifier that indicates the ordering of the configuration items of a resource.

$sel:version:BaseConfigurationItem', baseConfigurationItem_version - The version number of the resource configuration.

$sel:resourceType:BaseConfigurationItem', baseConfigurationItem_resourceType - The type of Amazon Web Services resource.

$sel:supplementaryConfiguration:BaseConfigurationItem', baseConfigurationItem_supplementaryConfiguration - Configuration attributes that Config returns for certain resource types to supplement the information returned for the configuration parameter.

$sel:availabilityZone:BaseConfigurationItem', baseConfigurationItem_availabilityZone - The Availability Zone associated with the resource.

$sel:configurationItemCaptureTime:BaseConfigurationItem', baseConfigurationItem_configurationItemCaptureTime - The time when the configuration recording was initiated.

$sel:configurationItemStatus:BaseConfigurationItem', baseConfigurationItem_configurationItemStatus - The configuration item status. The valid values are:

OK – The resource configuration has been updated

ResourceDiscovered – The resource was newly discovered

ResourceNotRecorded – The resource was discovered but its configuration was not recorded since the recorder excludes the recording of resources of this type

ResourceDeleted – The resource was deleted

ResourceDeletedNotRecorded – The resource was deleted but its configuration was not recorded since the recorder excludes the recording of resources of this type

The CIs do not incur any cost.

$sel:resourceCreationTime:BaseConfigurationItem', baseConfigurationItem_resourceCreationTime - The time stamp when the resource was created.

$sel:resourceName:BaseConfigurationItem', baseConfigurationItem_resourceName - The custom name of the resource, if available.

$sel:awsRegion:BaseConfigurationItem', baseConfigurationItem_awsRegion - The region where the resource resides.

baseConfigurationItem_resourceId :: Lens' BaseConfigurationItem (Maybe Text)

The ID of the resource (for example., sg-xxxxxx).

baseConfigurationItem_accountId :: Lens' BaseConfigurationItem (Maybe Text)

The 12-digit Amazon Web Services account ID associated with the resource.

baseConfigurationItem_configuration :: Lens' BaseConfigurationItem (Maybe Text)

The description of the resource configuration.

baseConfigurationItem_arn :: Lens' BaseConfigurationItem (Maybe Text)

The Amazon Resource Name (ARN) of the resource.

baseConfigurationItem_configurationStateId :: Lens' BaseConfigurationItem (Maybe Text)

An identifier that indicates the ordering of the configuration items of a resource.

baseConfigurationItem_version :: Lens' BaseConfigurationItem (Maybe Text)

The version number of the resource configuration.

baseConfigurationItem_resourceType :: Lens' BaseConfigurationItem (Maybe ResourceType)

The type of Amazon Web Services resource.

baseConfigurationItem_supplementaryConfiguration :: Lens' BaseConfigurationItem (Maybe (HashMap Text Text))

Configuration attributes that Config returns for certain resource types to supplement the information returned for the configuration parameter.

baseConfigurationItem_availabilityZone :: Lens' BaseConfigurationItem (Maybe Text)

The Availability Zone associated with the resource.

baseConfigurationItem_configurationItemCaptureTime :: Lens' BaseConfigurationItem (Maybe UTCTime)

The time when the configuration recording was initiated.

baseConfigurationItem_configurationItemStatus :: Lens' BaseConfigurationItem (Maybe ConfigurationItemStatus)

The configuration item status. The valid values are:

OK – The resource configuration has been updated

ResourceDiscovered – The resource was newly discovered

ResourceNotRecorded – The resource was discovered but its configuration was not recorded since the recorder excludes the recording of resources of this type

ResourceDeleted – The resource was deleted

ResourceDeletedNotRecorded – The resource was deleted but its configuration was not recorded since the recorder excludes the recording of resources of this type

The CIs do not incur any cost.

baseConfigurationItem_resourceCreationTime :: Lens' BaseConfigurationItem (Maybe UTCTime)

The time stamp when the resource was created.

baseConfigurationItem_resourceName :: Lens' BaseConfigurationItem (Maybe Text)

The custom name of the resource, if available.

baseConfigurationItem_awsRegion :: Lens' BaseConfigurationItem (Maybe Text)

The region where the resource resides.