Network.AWS.Config.Types.ComplianceSummary

Description

Documentation

data ComplianceSummary

The number of Config rules or Amazon Web Services resources that are compliant and noncompliant.

See: newComplianceSummary smart constructor.

Constructors

ComplianceSummary'

Fields

complianceSummaryTimestamp :: Maybe POSIX

The time that Config created the compliance summary.

nonCompliantResourceCount :: Maybe ComplianceContributorCount

The number of Config rules or Amazon Web Services resources that are noncompliant, up to a maximum of 25 for rules and 100 for resources.

compliantResourceCount :: Maybe ComplianceContributorCount

The number of Config rules or Amazon Web Services resources that are compliant, up to a maximum of 25 for rules and 100 for resources.

newComplianceSummary :: ComplianceSummary

Create a value of ComplianceSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:complianceSummaryTimestamp:ComplianceSummary', complianceSummary_complianceSummaryTimestamp - The time that Config created the compliance summary.

$sel:nonCompliantResourceCount:ComplianceSummary', complianceSummary_nonCompliantResourceCount - The number of Config rules or Amazon Web Services resources that are noncompliant, up to a maximum of 25 for rules and 100 for resources.

$sel:compliantResourceCount:ComplianceSummary', complianceSummary_compliantResourceCount - The number of Config rules or Amazon Web Services resources that are compliant, up to a maximum of 25 for rules and 100 for resources.

complianceSummary_complianceSummaryTimestamp :: Lens' ComplianceSummary (Maybe UTCTime)

The time that Config created the compliance summary.

complianceSummary_nonCompliantResourceCount :: Lens' ComplianceSummary (Maybe ComplianceContributorCount)

The number of Config rules or Amazon Web Services resources that are noncompliant, up to a maximum of 25 for rules and 100 for resources.

complianceSummary_compliantResourceCount :: Lens' ComplianceSummary (Maybe ComplianceContributorCount)

The number of Config rules or Amazon Web Services resources that are compliant, up to a maximum of 25 for rules and 100 for resources.