Network.AWS.Config.Types.ConfigRule

Description

Documentation

data ConfigRule

An Config rule represents an Lambda function that you create for a custom rule or a predefined function for an Config managed rule. The function evaluates configuration items to assess whether your Amazon Web Services resources comply with your desired configurations. This function can run when Config detects a configuration change to an Amazon Web Services resource and at a periodic frequency that you choose (for example, every 24 hours).

You can use the Amazon Web Services CLI and Amazon Web Services SDKs if you want to create a rule that triggers evaluations for your resources when Config delivers the configuration snapshot. For more information, see ConfigSnapshotDeliveryProperties.

For more information about developing and using Config rules, see Evaluating Amazon Web Services resource Configurations with Config in the Config Developer Guide.

See: newConfigRule smart constructor.

Constructors

ConfigRule'

Fields

configRuleArn :: Maybe Text

The Amazon Resource Name (ARN) of the Config rule.

configRuleId :: Maybe Text

The ID of the Config rule.

maximumExecutionFrequency :: Maybe MaximumExecutionFrequency

The maximum frequency with which Config runs evaluations for a rule. You can specify a value for MaximumExecutionFrequency when:

You are using an Config managed rule that is triggered at a periodic frequency.

Your custom rule is triggered when Config delivers the configuration snapshot. For more information, see ConfigSnapshotDeliveryProperties.

By default, rules with a periodic trigger are evaluated every 24 hours. To change the frequency, specify a valid value for the MaximumExecutionFrequency parameter.

configRuleName :: Maybe Text

The name that you assign to the Config rule. The name is required if you are adding a new rule.

configRuleState :: Maybe ConfigRuleState

Indicates whether the Config rule is active or is currently being deleted by Config. It can also indicate the evaluation status for the Config rule.

Config sets the state of the rule to EVALUATING temporarily after you use the StartConfigRulesEvaluation request to evaluate your resources against the Config rule.

Config sets the state of the rule to DELETING_RESULTS temporarily after you use the DeleteEvaluationResults request to delete the current evaluation results for the Config rule.

Config temporarily sets the state of a rule to DELETING after you use the DeleteConfigRule request to delete the rule. After Config deletes the rule, the rule and all of its evaluations are erased and are no longer available.

scope :: Maybe Scope

Defines which resources can trigger an evaluation for the rule. The scope can include one or more resource types, a combination of one resource type and one resource ID, or a combination of a tag key and value. Specify a scope to constrain the resources that can trigger an evaluation for the rule. If you do not specify a scope, evaluations are triggered when any resource in the recording group changes.

The scope can be empty.

inputParameters :: Maybe Text

A string, in JSON format, that is passed to the Config rule Lambda function.

description :: Maybe Text

The description that you provide for the Config rule.

createdBy :: Maybe Text

Service principal name of the service that created the rule.

The field is populated only if the service linked rule is created by a service. The field is empty if you create your own rule.

source :: Source

Provides the rule owner (Amazon Web Services or customer), the rule identifier, and the notifications that cause the function to evaluate your Amazon Web Services resources.

newConfigRule

Arguments

:: Source
$sel:source:ConfigRule'

-> ConfigRule

Create a value of ConfigRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configRuleArn:ConfigRule', configRule_configRuleArn - The Amazon Resource Name (ARN) of the Config rule.

$sel:configRuleId:ConfigRule', configRule_configRuleId - The ID of the Config rule.

$sel:maximumExecutionFrequency:ConfigRule', configRule_maximumExecutionFrequency - The maximum frequency with which Config runs evaluations for a rule. You can specify a value for MaximumExecutionFrequency when:

You are using an Config managed rule that is triggered at a periodic frequency.

Your custom rule is triggered when Config delivers the configuration snapshot. For more information, see ConfigSnapshotDeliveryProperties.

By default, rules with a periodic trigger are evaluated every 24 hours. To change the frequency, specify a valid value for the MaximumExecutionFrequency parameter.

$sel:configRuleName:ConfigRule', configRule_configRuleName - The name that you assign to the Config rule. The name is required if you are adding a new rule.

$sel:configRuleState:ConfigRule', configRule_configRuleState - Indicates whether the Config rule is active or is currently being deleted by Config. It can also indicate the evaluation status for the Config rule.

Config sets the state of the rule to EVALUATING temporarily after you use the StartConfigRulesEvaluation request to evaluate your resources against the Config rule.

Config sets the state of the rule to DELETING_RESULTS temporarily after you use the DeleteEvaluationResults request to delete the current evaluation results for the Config rule.

Config temporarily sets the state of a rule to DELETING after you use the DeleteConfigRule request to delete the rule. After Config deletes the rule, the rule and all of its evaluations are erased and are no longer available.

$sel:scope:ConfigRule', configRule_scope - Defines which resources can trigger an evaluation for the rule. The scope can include one or more resource types, a combination of one resource type and one resource ID, or a combination of a tag key and value. Specify a scope to constrain the resources that can trigger an evaluation for the rule. If you do not specify a scope, evaluations are triggered when any resource in the recording group changes.

The scope can be empty.

$sel:inputParameters:ConfigRule', configRule_inputParameters - A string, in JSON format, that is passed to the Config rule Lambda function.

$sel:description:ConfigRule', configRule_description - The description that you provide for the Config rule.

$sel:createdBy:ConfigRule', configRule_createdBy - Service principal name of the service that created the rule.

The field is populated only if the service linked rule is created by a service. The field is empty if you create your own rule.

$sel:source:ConfigRule', configRule_source - Provides the rule owner (Amazon Web Services or customer), the rule identifier, and the notifications that cause the function to evaluate your Amazon Web Services resources.

configRule_configRuleArn :: Lens' ConfigRule (Maybe Text)

The Amazon Resource Name (ARN) of the Config rule.

configRule_configRuleId :: Lens' ConfigRule (Maybe Text)

The ID of the Config rule.

configRule_maximumExecutionFrequency :: Lens' ConfigRule (Maybe MaximumExecutionFrequency)

The maximum frequency with which Config runs evaluations for a rule. You can specify a value for MaximumExecutionFrequency when:

You are using an Config managed rule that is triggered at a periodic frequency.

Your custom rule is triggered when Config delivers the configuration snapshot. For more information, see ConfigSnapshotDeliveryProperties.

By default, rules with a periodic trigger are evaluated every 24 hours. To change the frequency, specify a valid value for the MaximumExecutionFrequency parameter.

configRule_configRuleName :: Lens' ConfigRule (Maybe Text)

The name that you assign to the Config rule. The name is required if you are adding a new rule.

configRule_configRuleState :: Lens' ConfigRule (Maybe ConfigRuleState)

Indicates whether the Config rule is active or is currently being deleted by Config. It can also indicate the evaluation status for the Config rule.

Config sets the state of the rule to EVALUATING temporarily after you use the StartConfigRulesEvaluation request to evaluate your resources against the Config rule.

Config sets the state of the rule to DELETING_RESULTS temporarily after you use the DeleteEvaluationResults request to delete the current evaluation results for the Config rule.

Config temporarily sets the state of a rule to DELETING after you use the DeleteConfigRule request to delete the rule. After Config deletes the rule, the rule and all of its evaluations are erased and are no longer available.

configRule_scope :: Lens' ConfigRule (Maybe Scope)

Defines which resources can trigger an evaluation for the rule. The scope can include one or more resource types, a combination of one resource type and one resource ID, or a combination of a tag key and value. Specify a scope to constrain the resources that can trigger an evaluation for the rule. If you do not specify a scope, evaluations are triggered when any resource in the recording group changes.

The scope can be empty.

configRule_inputParameters :: Lens' ConfigRule (Maybe Text)

A string, in JSON format, that is passed to the Config rule Lambda function.

configRule_description :: Lens' ConfigRule (Maybe Text)

The description that you provide for the Config rule.

configRule_createdBy :: Lens' ConfigRule (Maybe Text)

Service principal name of the service that created the rule.

The field is populated only if the service linked rule is created by a service. The field is empty if you create your own rule.

configRule_source :: Lens' ConfigRule Source

Provides the rule owner (Amazon Web Services or customer), the rule identifier, and the notifications that cause the function to evaluate your Amazon Web Services resources.