Network.AWS.Config.Types.ConfigRuleEvaluationStatus

Description

Documentation

data ConfigRuleEvaluationStatus

Status information for your Config managed rules. The status includes information such as the last time the rule ran, the last time it failed, and the related error for the last failure.

This action does not return status information about custom Config rules.

See: newConfigRuleEvaluationStatus smart constructor.

Constructors

ConfigRuleEvaluationStatus'

Fields

lastErrorMessage :: Maybe Text

The error message that Config returned when the rule last failed.

configRuleArn :: Maybe Text

The Amazon Resource Name (ARN) of the Config rule.

configRuleId :: Maybe Text

The ID of the Config rule.

configRuleName :: Maybe Text

The name of the Config rule.

firstEvaluationStarted :: Maybe Bool

Indicates whether Config has evaluated your resources against the rule at least once.

true - Config has evaluated your Amazon Web Services resources against the rule at least once.

false - Config has not once finished evaluating your Amazon Web Services resources against the rule.

firstActivatedTime :: Maybe POSIX

The time that you first activated the Config rule.

lastFailedEvaluationTime :: Maybe POSIX

The time that Config last failed to evaluate your Amazon Web Services resources against the rule.

lastErrorCode :: Maybe Text

The error code that Config returned when the rule last failed.

lastFailedInvocationTime :: Maybe POSIX

The time that Config last failed to invoke the Config rule to evaluate your Amazon Web Services resources.

lastDeactivatedTime :: Maybe POSIX

The time that you last turned off the Config rule.

lastSuccessfulInvocationTime :: Maybe POSIX

The time that Config last successfully invoked the Config rule to evaluate your Amazon Web Services resources.

lastSuccessfulEvaluationTime :: Maybe POSIX

The time that Config last successfully evaluated your Amazon Web Services resources against the rule.

newConfigRuleEvaluationStatus :: ConfigRuleEvaluationStatus

Create a value of ConfigRuleEvaluationStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastErrorMessage:ConfigRuleEvaluationStatus', configRuleEvaluationStatus_lastErrorMessage - The error message that Config returned when the rule last failed.

$sel:configRuleArn:ConfigRuleEvaluationStatus', configRuleEvaluationStatus_configRuleArn - The Amazon Resource Name (ARN) of the Config rule.

$sel:configRuleId:ConfigRuleEvaluationStatus', configRuleEvaluationStatus_configRuleId - The ID of the Config rule.

$sel:configRuleName:ConfigRuleEvaluationStatus', configRuleEvaluationStatus_configRuleName - The name of the Config rule.

$sel:firstEvaluationStarted:ConfigRuleEvaluationStatus', configRuleEvaluationStatus_firstEvaluationStarted - Indicates whether Config has evaluated your resources against the rule at least once.

true - Config has evaluated your Amazon Web Services resources against the rule at least once.

false - Config has not once finished evaluating your Amazon Web Services resources against the rule.

$sel:firstActivatedTime:ConfigRuleEvaluationStatus', configRuleEvaluationStatus_firstActivatedTime - The time that you first activated the Config rule.

$sel:lastFailedEvaluationTime:ConfigRuleEvaluationStatus', configRuleEvaluationStatus_lastFailedEvaluationTime - The time that Config last failed to evaluate your Amazon Web Services resources against the rule.

$sel:lastErrorCode:ConfigRuleEvaluationStatus', configRuleEvaluationStatus_lastErrorCode - The error code that Config returned when the rule last failed.

$sel:lastFailedInvocationTime:ConfigRuleEvaluationStatus', configRuleEvaluationStatus_lastFailedInvocationTime - The time that Config last failed to invoke the Config rule to evaluate your Amazon Web Services resources.

$sel:lastDeactivatedTime:ConfigRuleEvaluationStatus', configRuleEvaluationStatus_lastDeactivatedTime - The time that you last turned off the Config rule.

$sel:lastSuccessfulInvocationTime:ConfigRuleEvaluationStatus', configRuleEvaluationStatus_lastSuccessfulInvocationTime - The time that Config last successfully invoked the Config rule to evaluate your Amazon Web Services resources.

$sel:lastSuccessfulEvaluationTime:ConfigRuleEvaluationStatus', configRuleEvaluationStatus_lastSuccessfulEvaluationTime - The time that Config last successfully evaluated your Amazon Web Services resources against the rule.

configRuleEvaluationStatus_lastErrorMessage :: Lens' ConfigRuleEvaluationStatus (Maybe Text)

The error message that Config returned when the rule last failed.

configRuleEvaluationStatus_configRuleArn :: Lens' ConfigRuleEvaluationStatus (Maybe Text)

The Amazon Resource Name (ARN) of the Config rule.

configRuleEvaluationStatus_configRuleId :: Lens' ConfigRuleEvaluationStatus (Maybe Text)

The ID of the Config rule.

configRuleEvaluationStatus_configRuleName :: Lens' ConfigRuleEvaluationStatus (Maybe Text)

The name of the Config rule.

configRuleEvaluationStatus_firstEvaluationStarted :: Lens' ConfigRuleEvaluationStatus (Maybe Bool)

Indicates whether Config has evaluated your resources against the rule at least once.

true - Config has evaluated your Amazon Web Services resources against the rule at least once.

false - Config has not once finished evaluating your Amazon Web Services resources against the rule.

configRuleEvaluationStatus_firstActivatedTime :: Lens' ConfigRuleEvaluationStatus (Maybe UTCTime)

The time that you first activated the Config rule.

configRuleEvaluationStatus_lastFailedEvaluationTime :: Lens' ConfigRuleEvaluationStatus (Maybe UTCTime)

The time that Config last failed to evaluate your Amazon Web Services resources against the rule.

configRuleEvaluationStatus_lastErrorCode :: Lens' ConfigRuleEvaluationStatus (Maybe Text)

The error code that Config returned when the rule last failed.

configRuleEvaluationStatus_lastFailedInvocationTime :: Lens' ConfigRuleEvaluationStatus (Maybe UTCTime)

The time that Config last failed to invoke the Config rule to evaluate your Amazon Web Services resources.

configRuleEvaluationStatus_lastDeactivatedTime :: Lens' ConfigRuleEvaluationStatus (Maybe UTCTime)

The time that you last turned off the Config rule.

configRuleEvaluationStatus_lastSuccessfulInvocationTime :: Lens' ConfigRuleEvaluationStatus (Maybe UTCTime)

The time that Config last successfully invoked the Config rule to evaluate your Amazon Web Services resources.

configRuleEvaluationStatus_lastSuccessfulEvaluationTime :: Lens' ConfigRuleEvaluationStatus (Maybe UTCTime)

The time that Config last successfully evaluated your Amazon Web Services resources against the rule.