Network.AWS.Config.Types.ConformancePackComplianceSummary

Description

Documentation

data ConformancePackComplianceSummary

Summary includes the name and status of the conformance pack.

See: newConformancePackComplianceSummary smart constructor.

Constructors

ConformancePackComplianceSummary'

Fields

conformancePackName :: Text

The name of the conformance pack name.

conformancePackComplianceStatus :: ConformancePackComplianceType

The status of the conformance pack. The allowed values are COMPLIANT, NON_COMPLIANT and INSUFFICIENT_DATA.

newConformancePackComplianceSummary

Arguments

:: Text
$sel:conformancePackName:ConformancePackComplianceSummary'

-> ConformancePackComplianceType
$sel:conformancePackComplianceStatus:ConformancePackComplianceSummary'

-> ConformancePackComplianceSummary

Create a value of ConformancePackComplianceSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:conformancePackName:ConformancePackComplianceSummary', conformancePackComplianceSummary_conformancePackName - The name of the conformance pack name.

$sel:conformancePackComplianceStatus:ConformancePackComplianceSummary', conformancePackComplianceSummary_conformancePackComplianceStatus - The status of the conformance pack. The allowed values are COMPLIANT, NON_COMPLIANT and INSUFFICIENT_DATA.

conformancePackComplianceSummary_conformancePackName :: Lens' ConformancePackComplianceSummary Text

The name of the conformance pack name.

conformancePackComplianceSummary_conformancePackComplianceStatus :: Lens' ConformancePackComplianceSummary ConformancePackComplianceType

The status of the conformance pack. The allowed values are COMPLIANT, NON_COMPLIANT and INSUFFICIENT_DATA.