Network.AWS.Config.Types.ConformancePackDetail

Description

Documentation

data ConformancePackDetail

Returns details of a conformance pack. A conformance pack is a collection of Config rules and remediation actions that can be easily deployed in an account and a region.

See: newConformancePackDetail smart constructor.

Constructors

ConformancePackDetail'

Fields

deliveryS3Bucket :: Maybe Text

The name of the Amazon S3 bucket where Config stores conformance pack templates.

This field is optional.

deliveryS3KeyPrefix :: Maybe Text

The prefix for the Amazon S3 bucket.

This field is optional.

createdBy :: Maybe Text

Amazon Web Services service that created the conformance pack.

lastUpdateRequestedTime :: Maybe POSIX

Last time when conformation pack update was requested.

conformancePackInputParameters :: Maybe [ConformancePackInputParameter]

A list of ConformancePackInputParameter objects.

conformancePackName :: Text

Name of the conformance pack.

conformancePackArn :: Text

Amazon Resource Name (ARN) of the conformance pack.

conformancePackId :: Text

ID of the conformance pack.

newConformancePackDetail

Arguments

:: Text
$sel:conformancePackName:ConformancePackDetail'

-> Text
$sel:conformancePackArn:ConformancePackDetail'

-> Text
$sel:conformancePackId:ConformancePackDetail'

-> ConformancePackDetail

Create a value of ConformancePackDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deliveryS3Bucket:ConformancePackDetail', conformancePackDetail_deliveryS3Bucket - The name of the Amazon S3 bucket where Config stores conformance pack templates.

This field is optional.

$sel:deliveryS3KeyPrefix:ConformancePackDetail', conformancePackDetail_deliveryS3KeyPrefix - The prefix for the Amazon S3 bucket.

This field is optional.

$sel:createdBy:ConformancePackDetail', conformancePackDetail_createdBy - Amazon Web Services service that created the conformance pack.

$sel:lastUpdateRequestedTime:ConformancePackDetail', conformancePackDetail_lastUpdateRequestedTime - Last time when conformation pack update was requested.

$sel:conformancePackInputParameters:ConformancePackDetail', conformancePackDetail_conformancePackInputParameters - A list of ConformancePackInputParameter objects.

$sel:conformancePackName:ConformancePackDetail', conformancePackDetail_conformancePackName - Name of the conformance pack.

$sel:conformancePackArn:ConformancePackDetail', conformancePackDetail_conformancePackArn - Amazon Resource Name (ARN) of the conformance pack.

$sel:conformancePackId:ConformancePackDetail', conformancePackDetail_conformancePackId - ID of the conformance pack.

conformancePackDetail_deliveryS3Bucket :: Lens' ConformancePackDetail (Maybe Text)

The name of the Amazon S3 bucket where Config stores conformance pack templates.

This field is optional.

conformancePackDetail_deliveryS3KeyPrefix :: Lens' ConformancePackDetail (Maybe Text)

The prefix for the Amazon S3 bucket.

This field is optional.

conformancePackDetail_createdBy :: Lens' ConformancePackDetail (Maybe Text)

Amazon Web Services service that created the conformance pack.

conformancePackDetail_lastUpdateRequestedTime :: Lens' ConformancePackDetail (Maybe UTCTime)

Last time when conformation pack update was requested.

conformancePackDetail_conformancePackInputParameters :: Lens' ConformancePackDetail (Maybe [ConformancePackInputParameter])

A list of ConformancePackInputParameter objects.

conformancePackDetail_conformancePackName :: Lens' ConformancePackDetail Text

Name of the conformance pack.

conformancePackDetail_conformancePackArn :: Lens' ConformancePackDetail Text

Amazon Resource Name (ARN) of the conformance pack.

conformancePackDetail_conformancePackId :: Lens' ConformancePackDetail Text

ID of the conformance pack.