Network.AWS.Config.Types.ConformancePackEvaluationResult

Description

Documentation

data ConformancePackEvaluationResult

The details of a conformance pack evaluation. Provides Config rule and Amazon Web Services resource type that was evaluated, the compliance of the conformance pack, related time stamps, and supplementary information.

See: newConformancePackEvaluationResult smart constructor.

Constructors

ConformancePackEvaluationResult'

Fields

annotation :: Maybe Text

Supplementary information about how the evaluation determined the compliance.

complianceType :: ConformancePackComplianceType

The compliance type. The allowed values are COMPLIANT and NON_COMPLIANT. INSUFFICIENT_DATA is not supported.

evaluationResultIdentifier :: EvaluationResultIdentifier

configRuleInvokedTime :: POSIX

The time when Config rule evaluated Amazon Web Services resource.

resultRecordedTime :: POSIX

The time when Config recorded the evaluation result.

newConformancePackEvaluationResult

Arguments

:: ConformancePackComplianceType
$sel:complianceType:ConformancePackEvaluationResult'

-> EvaluationResultIdentifier
$sel:evaluationResultIdentifier:ConformancePackEvaluationResult'

-> UTCTime
$sel:configRuleInvokedTime:ConformancePackEvaluationResult'

-> UTCTime
$sel:resultRecordedTime:ConformancePackEvaluationResult'

-> ConformancePackEvaluationResult

Create a value of ConformancePackEvaluationResult with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:annotation:ConformancePackEvaluationResult', conformancePackEvaluationResult_annotation - Supplementary information about how the evaluation determined the compliance.

$sel:complianceType:ConformancePackEvaluationResult', conformancePackEvaluationResult_complianceType - The compliance type. The allowed values are COMPLIANT and NON_COMPLIANT. INSUFFICIENT_DATA is not supported.

$sel:evaluationResultIdentifier:ConformancePackEvaluationResult', conformancePackEvaluationResult_evaluationResultIdentifier - Undocumented member.

$sel:configRuleInvokedTime:ConformancePackEvaluationResult', conformancePackEvaluationResult_configRuleInvokedTime - The time when Config rule evaluated Amazon Web Services resource.

$sel:resultRecordedTime:ConformancePackEvaluationResult', conformancePackEvaluationResult_resultRecordedTime - The time when Config recorded the evaluation result.

conformancePackEvaluationResult_annotation :: Lens' ConformancePackEvaluationResult (Maybe Text)

Supplementary information about how the evaluation determined the compliance.

conformancePackEvaluationResult_complianceType :: Lens' ConformancePackEvaluationResult ConformancePackComplianceType

The compliance type. The allowed values are COMPLIANT and NON_COMPLIANT. INSUFFICIENT_DATA is not supported.

conformancePackEvaluationResult_evaluationResultIdentifier :: Lens' ConformancePackEvaluationResult EvaluationResultIdentifier

Undocumented member.

conformancePackEvaluationResult_configRuleInvokedTime :: Lens' ConformancePackEvaluationResult UTCTime

The time when Config rule evaluated Amazon Web Services resource.

conformancePackEvaluationResult_resultRecordedTime :: Lens' ConformancePackEvaluationResult UTCTime

The time when Config recorded the evaluation result.