Network.AWS.Config.Types.DeliveryChannel

Description

Documentation

data DeliveryChannel

The channel through which Config delivers notifications and updated configuration states.

See: newDeliveryChannel smart constructor.

Constructors

DeliveryChannel'

Fields

s3KeyPrefix :: Maybe Text

The prefix for the specified Amazon S3 bucket.

snsTopicARN :: Maybe Text

The Amazon Resource Name (ARN) of the Amazon SNS topic to which Config sends notifications about configuration changes.

If you choose a topic from another account, the topic must have policies that grant access permissions to Config. For more information, see Permissions for the Amazon SNS Topic in the Config Developer Guide.

name :: Maybe Text

The name of the delivery channel. By default, Config assigns the name "default" when creating the delivery channel. To change the delivery channel name, you must use the DeleteDeliveryChannel action to delete your current delivery channel, and then you must use the PutDeliveryChannel command to create a delivery channel that has the desired name.

s3KmsKeyArn :: Maybe Text

The Amazon Resource Name (ARN) of the Key Management Service (KMS ) KMS key (KMS key) used to encrypt objects delivered by Config. Must belong to the same Region as the destination S3 bucket.

configSnapshotDeliveryProperties :: Maybe ConfigSnapshotDeliveryProperties

The options for how often Config delivers configuration snapshots to the Amazon S3 bucket.

s3BucketName :: Maybe Text

The name of the Amazon S3 bucket to which Config delivers configuration snapshots and configuration history files.

If you specify a bucket that belongs to another Amazon Web Services account, that bucket must have policies that grant access permissions to Config. For more information, see Permissions for the Amazon S3 Bucket in the Config Developer Guide.

newDeliveryChannel :: DeliveryChannel

Create a value of DeliveryChannel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3KeyPrefix:DeliveryChannel', deliveryChannel_s3KeyPrefix - The prefix for the specified Amazon S3 bucket.

$sel:snsTopicARN:DeliveryChannel', deliveryChannel_snsTopicARN - The Amazon Resource Name (ARN) of the Amazon SNS topic to which Config sends notifications about configuration changes.

If you choose a topic from another account, the topic must have policies that grant access permissions to Config. For more information, see Permissions for the Amazon SNS Topic in the Config Developer Guide.

$sel:name:DeliveryChannel', deliveryChannel_name - The name of the delivery channel. By default, Config assigns the name "default" when creating the delivery channel. To change the delivery channel name, you must use the DeleteDeliveryChannel action to delete your current delivery channel, and then you must use the PutDeliveryChannel command to create a delivery channel that has the desired name.

$sel:s3KmsKeyArn:DeliveryChannel', deliveryChannel_s3KmsKeyArn - The Amazon Resource Name (ARN) of the Key Management Service (KMS ) KMS key (KMS key) used to encrypt objects delivered by Config. Must belong to the same Region as the destination S3 bucket.

$sel:configSnapshotDeliveryProperties:DeliveryChannel', deliveryChannel_configSnapshotDeliveryProperties - The options for how often Config delivers configuration snapshots to the Amazon S3 bucket.

$sel:s3BucketName:DeliveryChannel', deliveryChannel_s3BucketName - The name of the Amazon S3 bucket to which Config delivers configuration snapshots and configuration history files.

If you specify a bucket that belongs to another Amazon Web Services account, that bucket must have policies that grant access permissions to Config. For more information, see Permissions for the Amazon S3 Bucket in the Config Developer Guide.

deliveryChannel_s3KeyPrefix :: Lens' DeliveryChannel (Maybe Text)

The prefix for the specified Amazon S3 bucket.

deliveryChannel_snsTopicARN :: Lens' DeliveryChannel (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon SNS topic to which Config sends notifications about configuration changes.

If you choose a topic from another account, the topic must have policies that grant access permissions to Config. For more information, see Permissions for the Amazon SNS Topic in the Config Developer Guide.

deliveryChannel_name :: Lens' DeliveryChannel (Maybe Text)

The name of the delivery channel. By default, Config assigns the name "default" when creating the delivery channel. To change the delivery channel name, you must use the DeleteDeliveryChannel action to delete your current delivery channel, and then you must use the PutDeliveryChannel command to create a delivery channel that has the desired name.

deliveryChannel_s3KmsKeyArn :: Lens' DeliveryChannel (Maybe Text)

The Amazon Resource Name (ARN) of the Key Management Service (KMS ) KMS key (KMS key) used to encrypt objects delivered by Config. Must belong to the same Region as the destination S3 bucket.

deliveryChannel_configSnapshotDeliveryProperties :: Lens' DeliveryChannel (Maybe ConfigSnapshotDeliveryProperties)

The options for how often Config delivers configuration snapshots to the Amazon S3 bucket.

deliveryChannel_s3BucketName :: Lens' DeliveryChannel (Maybe Text)

The name of the Amazon S3 bucket to which Config delivers configuration snapshots and configuration history files.

If you specify a bucket that belongs to another Amazon Web Services account, that bucket must have policies that grant access permissions to Config. For more information, see Permissions for the Amazon S3 Bucket in the Config Developer Guide.