Network.AWS.Config.Types.EvaluationResult

Description

Documentation

data EvaluationResult

The details of an Config evaluation. Provides the Amazon Web Services resource that was evaluated, the compliance of the resource, related time stamps, and supplementary information.

See: newEvaluationResult smart constructor.

Constructors

EvaluationResult'

Fields

annotation :: Maybe Text

Supplementary information about how the evaluation determined the compliance.

evaluationResultIdentifier :: Maybe EvaluationResultIdentifier

Uniquely identifies the evaluation result.

resultRecordedTime :: Maybe POSIX

The time when Config recorded the evaluation result.

complianceType :: Maybe ComplianceType

Indicates whether the Amazon Web Services resource complies with the Config rule that evaluated it.

For the EvaluationResult data type, Config supports only the COMPLIANT, NON_COMPLIANT, and NOT_APPLICABLE values. Config does not support the INSUFFICIENT_DATA value for the EvaluationResult data type.

configRuleInvokedTime :: Maybe POSIX

The time when the Config rule evaluated the Amazon Web Services resource.

resultToken :: Maybe Text

An encrypted token that associates an evaluation with an Config rule. The token identifies the rule, the Amazon Web Services resource being evaluated, and the event that triggered the evaluation.

newEvaluationResult :: EvaluationResult

Create a value of EvaluationResult with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:annotation:EvaluationResult', evaluationResult_annotation - Supplementary information about how the evaluation determined the compliance.

$sel:evaluationResultIdentifier:EvaluationResult', evaluationResult_evaluationResultIdentifier - Uniquely identifies the evaluation result.

$sel:resultRecordedTime:EvaluationResult', evaluationResult_resultRecordedTime - The time when Config recorded the evaluation result.

$sel:complianceType:EvaluationResult', evaluationResult_complianceType - Indicates whether the Amazon Web Services resource complies with the Config rule that evaluated it.

For the EvaluationResult data type, Config supports only the COMPLIANT, NON_COMPLIANT, and NOT_APPLICABLE values. Config does not support the INSUFFICIENT_DATA value for the EvaluationResult data type.

$sel:configRuleInvokedTime:EvaluationResult', evaluationResult_configRuleInvokedTime - The time when the Config rule evaluated the Amazon Web Services resource.

$sel:resultToken:EvaluationResult', evaluationResult_resultToken - An encrypted token that associates an evaluation with an Config rule. The token identifies the rule, the Amazon Web Services resource being evaluated, and the event that triggered the evaluation.

evaluationResult_annotation :: Lens' EvaluationResult (Maybe Text)

Supplementary information about how the evaluation determined the compliance.

evaluationResult_evaluationResultIdentifier :: Lens' EvaluationResult (Maybe EvaluationResultIdentifier)

Uniquely identifies the evaluation result.

evaluationResult_resultRecordedTime :: Lens' EvaluationResult (Maybe UTCTime)

The time when Config recorded the evaluation result.

evaluationResult_complianceType :: Lens' EvaluationResult (Maybe ComplianceType)

Indicates whether the Amazon Web Services resource complies with the Config rule that evaluated it.

For the EvaluationResult data type, Config supports only the COMPLIANT, NON_COMPLIANT, and NOT_APPLICABLE values. Config does not support the INSUFFICIENT_DATA value for the EvaluationResult data type.

evaluationResult_configRuleInvokedTime :: Lens' EvaluationResult (Maybe UTCTime)

The time when the Config rule evaluated the Amazon Web Services resource.

evaluationResult_resultToken :: Lens' EvaluationResult (Maybe Text)

An encrypted token that associates an evaluation with an Config rule. The token identifies the rule, the Amazon Web Services resource being evaluated, and the event that triggered the evaluation.