Network.AWS.Config.Types.ExternalEvaluation

Description

Documentation

data ExternalEvaluation

Identifies an Amazon Web Services resource and indicates whether it complies with the Config rule that it was evaluated against.

See: newExternalEvaluation smart constructor.

Constructors

ExternalEvaluation'

Fields

annotation :: Maybe Text

Supplementary information about the reason of compliance. For example, this task was completed on a specific date.

complianceResourceType :: Text

The evaluated compliance resource type. Config accepts AWS::::Account resource type.

complianceResourceId :: Text

The evaluated compliance resource ID. Config accepts only Amazon Web Services account ID.

complianceType :: ComplianceType

The compliance of the Amazon Web Services resource. The valid values are COMPLIANT, NON_COMPLIANT, and NOT_APPLICABLE.

orderingTimestamp :: POSIX

The time when the compliance was recorded.

newExternalEvaluation

Arguments

:: Text
$sel:complianceResourceType:ExternalEvaluation'

-> Text
$sel:complianceResourceId:ExternalEvaluation'

-> ComplianceType
$sel:complianceType:ExternalEvaluation'

-> UTCTime
$sel:orderingTimestamp:ExternalEvaluation'

-> ExternalEvaluation

Create a value of ExternalEvaluation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:annotation:ExternalEvaluation', externalEvaluation_annotation - Supplementary information about the reason of compliance. For example, this task was completed on a specific date.

$sel:complianceResourceType:ExternalEvaluation', externalEvaluation_complianceResourceType - The evaluated compliance resource type. Config accepts AWS::::Account resource type.

$sel:complianceResourceId:ExternalEvaluation', externalEvaluation_complianceResourceId - The evaluated compliance resource ID. Config accepts only Amazon Web Services account ID.

$sel:complianceType:ExternalEvaluation', externalEvaluation_complianceType - The compliance of the Amazon Web Services resource. The valid values are COMPLIANT, NON_COMPLIANT, and NOT_APPLICABLE.

$sel:orderingTimestamp:ExternalEvaluation', externalEvaluation_orderingTimestamp - The time when the compliance was recorded.

externalEvaluation_annotation :: Lens' ExternalEvaluation (Maybe Text)

Supplementary information about the reason of compliance. For example, this task was completed on a specific date.

externalEvaluation_complianceResourceType :: Lens' ExternalEvaluation Text

The evaluated compliance resource type. Config accepts AWS::::Account resource type.

externalEvaluation_complianceResourceId :: Lens' ExternalEvaluation Text

The evaluated compliance resource ID. Config accepts only Amazon Web Services account ID.

externalEvaluation_complianceType :: Lens' ExternalEvaluation ComplianceType

The compliance of the Amazon Web Services resource. The valid values are COMPLIANT, NON_COMPLIANT, and NOT_APPLICABLE.

externalEvaluation_orderingTimestamp :: Lens' ExternalEvaluation UTCTime

The time when the compliance was recorded.