Network.AWS.Config.Types.FailedDeleteRemediationExceptionsBatch

Description

Documentation

data FailedDeleteRemediationExceptionsBatch

List of each of the failed delete remediation exceptions with specific reasons.

See: newFailedDeleteRemediationExceptionsBatch smart constructor.

Constructors

FailedDeleteRemediationExceptionsBatch'

Fields

failureMessage :: Maybe Text

Returns a failure message for delete remediation exception. For example, Config creates an exception due to an internal error.

failedItems :: Maybe (NonEmpty RemediationExceptionResourceKey)

Returns remediation exception resource key object of the failed items.

newFailedDeleteRemediationExceptionsBatch :: FailedDeleteRemediationExceptionsBatch

Create a value of FailedDeleteRemediationExceptionsBatch with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureMessage:FailedDeleteRemediationExceptionsBatch', failedDeleteRemediationExceptionsBatch_failureMessage - Returns a failure message for delete remediation exception. For example, Config creates an exception due to an internal error.

$sel:failedItems:FailedDeleteRemediationExceptionsBatch', failedDeleteRemediationExceptionsBatch_failedItems - Returns remediation exception resource key object of the failed items.

failedDeleteRemediationExceptionsBatch_failureMessage :: Lens' FailedDeleteRemediationExceptionsBatch (Maybe Text)

Returns a failure message for delete remediation exception. For example, Config creates an exception due to an internal error.

failedDeleteRemediationExceptionsBatch_failedItems :: Lens' FailedDeleteRemediationExceptionsBatch (Maybe (NonEmpty RemediationExceptionResourceKey))

Returns remediation exception resource key object of the failed items.