Network.AWS.Config.Types.MemberAccountStatus

Description

Documentation

data MemberAccountStatus

Organization config rule creation or deletion status in each member account. This includes the name of the rule, the status, error code and error message when the rule creation or deletion failed.

See: newMemberAccountStatus smart constructor.

Constructors

MemberAccountStatus'

Fields

lastUpdateTime :: Maybe POSIX

The timestamp of the last status update.

errorMessage :: Maybe Text

An error message indicating that config rule account creation or deletion has failed due to an error in the member account.

errorCode :: Maybe Text

An error code that is returned when config rule creation or deletion failed in the member account.

accountId :: Text

The 12-digit account ID of a member account.

configRuleName :: Text

The name of config rule deployed in the member account.

memberAccountRuleStatus :: MemberAccountRuleStatus

Indicates deployment status for config rule in the member account. When master account calls PutOrganizationConfigRule action for the first time, config rule status is created in the member account. When master account calls PutOrganizationConfigRule action for the second time, config rule status is updated in the member account. Config rule status is deleted when the master account deletes OrganizationConfigRule and disables service access for config-multiaccountsetup.amazonaws.com.

Config sets the state of the rule to:

CREATE_SUCCESSFUL when config rule has been created in the member account.

CREATE_IN_PROGRESS when config rule is being created in the member account.

CREATE_FAILED when config rule creation has failed in the member account.

DELETE_FAILED when config rule deletion has failed in the member account.

DELETE_IN_PROGRESS when config rule is being deleted in the member account.

DELETE_SUCCESSFUL when config rule has been deleted in the member account.

UPDATE_SUCCESSFUL when config rule has been updated in the member account.

UPDATE_IN_PROGRESS when config rule is being updated in the member account.

UPDATE_FAILED when config rule deletion has failed in the member account.

newMemberAccountStatus

Arguments

:: Text
$sel:accountId:MemberAccountStatus'

-> Text
$sel:configRuleName:MemberAccountStatus'

-> MemberAccountRuleStatus
$sel:memberAccountRuleStatus:MemberAccountStatus'

-> MemberAccountStatus

Create a value of MemberAccountStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastUpdateTime:MemberAccountStatus', memberAccountStatus_lastUpdateTime - The timestamp of the last status update.

$sel:errorMessage:MemberAccountStatus', memberAccountStatus_errorMessage - An error message indicating that config rule account creation or deletion has failed due to an error in the member account.

$sel:errorCode:MemberAccountStatus', memberAccountStatus_errorCode - An error code that is returned when config rule creation or deletion failed in the member account.

$sel:accountId:MemberAccountStatus', memberAccountStatus_accountId - The 12-digit account ID of a member account.

$sel:configRuleName:MemberAccountStatus', memberAccountStatus_configRuleName - The name of config rule deployed in the member account.

$sel:memberAccountRuleStatus:MemberAccountStatus', memberAccountStatus_memberAccountRuleStatus - Indicates deployment status for config rule in the member account. When master account calls PutOrganizationConfigRule action for the first time, config rule status is created in the member account. When master account calls PutOrganizationConfigRule action for the second time, config rule status is updated in the member account. Config rule status is deleted when the master account deletes OrganizationConfigRule and disables service access for config-multiaccountsetup.amazonaws.com.

Config sets the state of the rule to:

CREATE_SUCCESSFUL when config rule has been created in the member account.

CREATE_IN_PROGRESS when config rule is being created in the member account.

CREATE_FAILED when config rule creation has failed in the member account.

DELETE_FAILED when config rule deletion has failed in the member account.

DELETE_IN_PROGRESS when config rule is being deleted in the member account.

DELETE_SUCCESSFUL when config rule has been deleted in the member account.

UPDATE_SUCCESSFUL when config rule has been updated in the member account.

UPDATE_IN_PROGRESS when config rule is being updated in the member account.

UPDATE_FAILED when config rule deletion has failed in the member account.

memberAccountStatus_lastUpdateTime :: Lens' MemberAccountStatus (Maybe UTCTime)

The timestamp of the last status update.

memberAccountStatus_errorMessage :: Lens' MemberAccountStatus (Maybe Text)

An error message indicating that config rule account creation or deletion has failed due to an error in the member account.

memberAccountStatus_errorCode :: Lens' MemberAccountStatus (Maybe Text)

An error code that is returned when config rule creation or deletion failed in the member account.

memberAccountStatus_accountId :: Lens' MemberAccountStatus Text

The 12-digit account ID of a member account.

memberAccountStatus_configRuleName :: Lens' MemberAccountStatus Text

The name of config rule deployed in the member account.

memberAccountStatus_memberAccountRuleStatus :: Lens' MemberAccountStatus MemberAccountRuleStatus

Indicates deployment status for config rule in the member account. When master account calls PutOrganizationConfigRule action for the first time, config rule status is created in the member account. When master account calls PutOrganizationConfigRule action for the second time, config rule status is updated in the member account. Config rule status is deleted when the master account deletes OrganizationConfigRule and disables service access for config-multiaccountsetup.amazonaws.com.

Config sets the state of the rule to:

CREATE_SUCCESSFUL when config rule has been created in the member account.

CREATE_IN_PROGRESS when config rule is being created in the member account.

CREATE_FAILED when config rule creation has failed in the member account.

DELETE_FAILED when config rule deletion has failed in the member account.

DELETE_IN_PROGRESS when config rule is being deleted in the member account.

DELETE_SUCCESSFUL when config rule has been deleted in the member account.

UPDATE_SUCCESSFUL when config rule has been updated in the member account.

UPDATE_IN_PROGRESS when config rule is being updated in the member account.

UPDATE_FAILED when config rule deletion has failed in the member account.