Network.AWS.Config.Types.OrganizationConfigRule

Description

Documentation

data OrganizationConfigRule

An organization config rule that has information about config rules that Config creates in member accounts.

See: newOrganizationConfigRule smart constructor.

Constructors

OrganizationConfigRule'

Fields

lastUpdateTime :: Maybe POSIX

The timestamp of the last update.

organizationManagedRuleMetadata :: Maybe OrganizationManagedRuleMetadata

An OrganizationManagedRuleMetadata object.

organizationCustomRuleMetadata :: Maybe OrganizationCustomRuleMetadata

An OrganizationCustomRuleMetadata object.

excludedAccounts :: Maybe [Text]

A comma-separated list of accounts excluded from organization config rule.

organizationConfigRuleName :: Text

The name that you assign to organization config rule.

organizationConfigRuleArn :: Text

Amazon Resource Name (ARN) of organization config rule.

newOrganizationConfigRule

Arguments

:: Text
$sel:organizationConfigRuleName:OrganizationConfigRule'

-> Text
$sel:organizationConfigRuleArn:OrganizationConfigRule'

-> OrganizationConfigRule

Create a value of OrganizationConfigRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastUpdateTime:OrganizationConfigRule', organizationConfigRule_lastUpdateTime - The timestamp of the last update.

$sel:organizationManagedRuleMetadata:OrganizationConfigRule', organizationConfigRule_organizationManagedRuleMetadata - An OrganizationManagedRuleMetadata object.

$sel:organizationCustomRuleMetadata:OrganizationConfigRule', organizationConfigRule_organizationCustomRuleMetadata - An OrganizationCustomRuleMetadata object.

$sel:excludedAccounts:OrganizationConfigRule', organizationConfigRule_excludedAccounts - A comma-separated list of accounts excluded from organization config rule.

$sel:organizationConfigRuleName:OrganizationConfigRule', organizationConfigRule_organizationConfigRuleName - The name that you assign to organization config rule.

$sel:organizationConfigRuleArn:OrganizationConfigRule', organizationConfigRule_organizationConfigRuleArn - Amazon Resource Name (ARN) of organization config rule.

organizationConfigRule_lastUpdateTime :: Lens' OrganizationConfigRule (Maybe UTCTime)

The timestamp of the last update.

organizationConfigRule_organizationManagedRuleMetadata :: Lens' OrganizationConfigRule (Maybe OrganizationManagedRuleMetadata)

An OrganizationManagedRuleMetadata object.

organizationConfigRule_organizationCustomRuleMetadata :: Lens' OrganizationConfigRule (Maybe OrganizationCustomRuleMetadata)

An OrganizationCustomRuleMetadata object.

organizationConfigRule_excludedAccounts :: Lens' OrganizationConfigRule (Maybe [Text])

A comma-separated list of accounts excluded from organization config rule.

organizationConfigRule_organizationConfigRuleName :: Lens' OrganizationConfigRule Text

The name that you assign to organization config rule.

organizationConfigRule_organizationConfigRuleArn :: Lens' OrganizationConfigRule Text

Amazon Resource Name (ARN) of organization config rule.