Network.AWS.Config.Types.OrganizationConfigRuleStatus

Description

Documentation

data OrganizationConfigRuleStatus

Returns the status for an organization config rule in an organization.

See: newOrganizationConfigRuleStatus smart constructor.

Constructors

OrganizationConfigRuleStatus'

Fields

lastUpdateTime :: Maybe POSIX

The timestamp of the last update.

errorMessage :: Maybe Text

An error message indicating that organization config rule creation or deletion failed due to an error.

errorCode :: Maybe Text

An error code that is returned when organization config rule creation or deletion has failed.

organizationConfigRuleName :: Text

The name that you assign to organization config rule.

organizationRuleStatus :: OrganizationRuleStatus

Indicates deployment status of an organization config rule. When master account calls PutOrganizationConfigRule action for the first time, config rule status is created in all the member accounts. When master account calls PutOrganizationConfigRule action for the second time, config rule status is updated in all the member accounts. Additionally, config rule status is updated when one or more member accounts join or leave an organization. Config rule status is deleted when the master account deletes OrganizationConfigRule in all the member accounts and disables service access for config-multiaccountsetup.amazonaws.com.

Config sets the state of the rule to:

CREATE_SUCCESSFUL when an organization config rule has been successfully created in all the member accounts.

CREATE_IN_PROGRESS when an organization config rule creation is in progress.

CREATE_FAILED when an organization config rule creation failed in one or more member accounts within that organization.

DELETE_FAILED when an organization config rule deletion failed in one or more member accounts within that organization.

DELETE_IN_PROGRESS when an organization config rule deletion is in progress.

DELETE_SUCCESSFUL when an organization config rule has been successfully deleted from all the member accounts.

UPDATE_SUCCESSFUL when an organization config rule has been successfully updated in all the member accounts.

UPDATE_IN_PROGRESS when an organization config rule update is in progress.

UPDATE_FAILED when an organization config rule update failed in one or more member accounts within that organization.

newOrganizationConfigRuleStatus

Arguments

:: Text
$sel:organizationConfigRuleName:OrganizationConfigRuleStatus'

-> OrganizationRuleStatus
$sel:organizationRuleStatus:OrganizationConfigRuleStatus'

-> OrganizationConfigRuleStatus

Create a value of OrganizationConfigRuleStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastUpdateTime:OrganizationConfigRuleStatus', organizationConfigRuleStatus_lastUpdateTime - The timestamp of the last update.

$sel:errorMessage:OrganizationConfigRuleStatus', organizationConfigRuleStatus_errorMessage - An error message indicating that organization config rule creation or deletion failed due to an error.

$sel:errorCode:OrganizationConfigRuleStatus', organizationConfigRuleStatus_errorCode - An error code that is returned when organization config rule creation or deletion has failed.

$sel:organizationConfigRuleName:OrganizationConfigRuleStatus', organizationConfigRuleStatus_organizationConfigRuleName - The name that you assign to organization config rule.

$sel:organizationRuleStatus:OrganizationConfigRuleStatus', organizationConfigRuleStatus_organizationRuleStatus - Indicates deployment status of an organization config rule. When master account calls PutOrganizationConfigRule action for the first time, config rule status is created in all the member accounts. When master account calls PutOrganizationConfigRule action for the second time, config rule status is updated in all the member accounts. Additionally, config rule status is updated when one or more member accounts join or leave an organization. Config rule status is deleted when the master account deletes OrganizationConfigRule in all the member accounts and disables service access for config-multiaccountsetup.amazonaws.com.

Config sets the state of the rule to:

CREATE_SUCCESSFUL when an organization config rule has been successfully created in all the member accounts.

CREATE_IN_PROGRESS when an organization config rule creation is in progress.

CREATE_FAILED when an organization config rule creation failed in one or more member accounts within that organization.

DELETE_FAILED when an organization config rule deletion failed in one or more member accounts within that organization.

DELETE_IN_PROGRESS when an organization config rule deletion is in progress.

DELETE_SUCCESSFUL when an organization config rule has been successfully deleted from all the member accounts.

UPDATE_SUCCESSFUL when an organization config rule has been successfully updated in all the member accounts.

UPDATE_IN_PROGRESS when an organization config rule update is in progress.

UPDATE_FAILED when an organization config rule update failed in one or more member accounts within that organization.

organizationConfigRuleStatus_lastUpdateTime :: Lens' OrganizationConfigRuleStatus (Maybe UTCTime)

The timestamp of the last update.

organizationConfigRuleStatus_errorMessage :: Lens' OrganizationConfigRuleStatus (Maybe Text)

An error message indicating that organization config rule creation or deletion failed due to an error.

organizationConfigRuleStatus_errorCode :: Lens' OrganizationConfigRuleStatus (Maybe Text)

An error code that is returned when organization config rule creation or deletion has failed.

organizationConfigRuleStatus_organizationConfigRuleName :: Lens' OrganizationConfigRuleStatus Text

The name that you assign to organization config rule.

organizationConfigRuleStatus_organizationRuleStatus :: Lens' OrganizationConfigRuleStatus OrganizationRuleStatus

Indicates deployment status of an organization config rule. When master account calls PutOrganizationConfigRule action for the first time, config rule status is created in all the member accounts. When master account calls PutOrganizationConfigRule action for the second time, config rule status is updated in all the member accounts. Additionally, config rule status is updated when one or more member accounts join or leave an organization. Config rule status is deleted when the master account deletes OrganizationConfigRule in all the member accounts and disables service access for config-multiaccountsetup.amazonaws.com.

Config sets the state of the rule to:

CREATE_SUCCESSFUL when an organization config rule has been successfully created in all the member accounts.

CREATE_IN_PROGRESS when an organization config rule creation is in progress.

CREATE_FAILED when an organization config rule creation failed in one or more member accounts within that organization.

DELETE_FAILED when an organization config rule deletion failed in one or more member accounts within that organization.

DELETE_IN_PROGRESS when an organization config rule deletion is in progress.

DELETE_SUCCESSFUL when an organization config rule has been successfully deleted from all the member accounts.

UPDATE_SUCCESSFUL when an organization config rule has been successfully updated in all the member accounts.

UPDATE_IN_PROGRESS when an organization config rule update is in progress.

UPDATE_FAILED when an organization config rule update failed in one or more member accounts within that organization.