Network.AWS.Config.Types.OrganizationCustomRuleMetadata

Description

Documentation

data OrganizationCustomRuleMetadata

An object that specifies organization custom rule metadata such as resource type, resource ID of Amazon Web Services resource, Lambda function ARN, and organization trigger types that trigger Config to evaluate your Amazon Web Services resources against a rule. It also provides the frequency with which you want Config to run evaluations for the rule if the trigger type is periodic.

See: newOrganizationCustomRuleMetadata smart constructor.

Constructors

OrganizationCustomRuleMetadata'

Fields

tagKeyScope :: Maybe Text

One part of a key-value pair that make up a tag. A key is a general label that acts like a category for more specific tag values.

maximumExecutionFrequency :: Maybe MaximumExecutionFrequency

The maximum frequency with which Config runs evaluations for a rule. Your custom rule is triggered when Config delivers the configuration snapshot. For more information, see ConfigSnapshotDeliveryProperties.

By default, rules with a periodic trigger are evaluated every 24 hours. To change the frequency, specify a valid value for the MaximumExecutionFrequency parameter.

resourceIdScope :: Maybe Text

The ID of the Amazon Web Services resource that was evaluated.

inputParameters :: Maybe Text

A string, in JSON format, that is passed to organization config rule Lambda function.

description :: Maybe Text

The description that you provide for organization config rule.

resourceTypesScope :: Maybe [Text]

The type of the Amazon Web Services resource that was evaluated.

tagValueScope :: Maybe Text

The optional part of a key-value pair that make up a tag. A value acts as a descriptor within a tag category (key).

lambdaFunctionArn :: Text

The lambda function ARN.

organizationConfigRuleTriggerTypes :: [OrganizationConfigRuleTriggerType]

The type of notification that triggers Config to run an evaluation for a rule. You can specify the following notification types:

ConfigurationItemChangeNotification - Triggers an evaluation when Config delivers a configuration item as a result of a resource change.

OversizedConfigurationItemChangeNotification - Triggers an evaluation when Config delivers an oversized configuration item. Config may generate this notification type when a resource changes and the notification exceeds the maximum size allowed by Amazon SNS.

ScheduledNotification - Triggers a periodic evaluation at the frequency specified for MaximumExecutionFrequency.

newOrganizationCustomRuleMetadata

Arguments

:: Text
$sel:lambdaFunctionArn:OrganizationCustomRuleMetadata'

-> OrganizationCustomRuleMetadata

Create a value of OrganizationCustomRuleMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tagKeyScope:OrganizationCustomRuleMetadata', organizationCustomRuleMetadata_tagKeyScope - One part of a key-value pair that make up a tag. A key is a general label that acts like a category for more specific tag values.

$sel:maximumExecutionFrequency:OrganizationCustomRuleMetadata', organizationCustomRuleMetadata_maximumExecutionFrequency - The maximum frequency with which Config runs evaluations for a rule. Your custom rule is triggered when Config delivers the configuration snapshot. For more information, see ConfigSnapshotDeliveryProperties.

By default, rules with a periodic trigger are evaluated every 24 hours. To change the frequency, specify a valid value for the MaximumExecutionFrequency parameter.

$sel:resourceIdScope:OrganizationCustomRuleMetadata', organizationCustomRuleMetadata_resourceIdScope - The ID of the Amazon Web Services resource that was evaluated.

$sel:inputParameters:OrganizationCustomRuleMetadata', organizationCustomRuleMetadata_inputParameters - A string, in JSON format, that is passed to organization config rule Lambda function.

$sel:description:OrganizationCustomRuleMetadata', organizationCustomRuleMetadata_description - The description that you provide for organization config rule.

$sel:resourceTypesScope:OrganizationCustomRuleMetadata', organizationCustomRuleMetadata_resourceTypesScope - The type of the Amazon Web Services resource that was evaluated.

$sel:tagValueScope:OrganizationCustomRuleMetadata', organizationCustomRuleMetadata_tagValueScope - The optional part of a key-value pair that make up a tag. A value acts as a descriptor within a tag category (key).

$sel:lambdaFunctionArn:OrganizationCustomRuleMetadata', organizationCustomRuleMetadata_lambdaFunctionArn - The lambda function ARN.

$sel:organizationConfigRuleTriggerTypes:OrganizationCustomRuleMetadata', organizationCustomRuleMetadata_organizationConfigRuleTriggerTypes - The type of notification that triggers Config to run an evaluation for a rule. You can specify the following notification types:

ConfigurationItemChangeNotification - Triggers an evaluation when Config delivers a configuration item as a result of a resource change.

OversizedConfigurationItemChangeNotification - Triggers an evaluation when Config delivers an oversized configuration item. Config may generate this notification type when a resource changes and the notification exceeds the maximum size allowed by Amazon SNS.

ScheduledNotification - Triggers a periodic evaluation at the frequency specified for MaximumExecutionFrequency.

organizationCustomRuleMetadata_tagKeyScope :: Lens' OrganizationCustomRuleMetadata (Maybe Text)

One part of a key-value pair that make up a tag. A key is a general label that acts like a category for more specific tag values.

organizationCustomRuleMetadata_maximumExecutionFrequency :: Lens' OrganizationCustomRuleMetadata (Maybe MaximumExecutionFrequency)

The maximum frequency with which Config runs evaluations for a rule. Your custom rule is triggered when Config delivers the configuration snapshot. For more information, see ConfigSnapshotDeliveryProperties.

By default, rules with a periodic trigger are evaluated every 24 hours. To change the frequency, specify a valid value for the MaximumExecutionFrequency parameter.

organizationCustomRuleMetadata_resourceIdScope :: Lens' OrganizationCustomRuleMetadata (Maybe Text)

The ID of the Amazon Web Services resource that was evaluated.

organizationCustomRuleMetadata_inputParameters :: Lens' OrganizationCustomRuleMetadata (Maybe Text)

A string, in JSON format, that is passed to organization config rule Lambda function.

organizationCustomRuleMetadata_description :: Lens' OrganizationCustomRuleMetadata (Maybe Text)

The description that you provide for organization config rule.

organizationCustomRuleMetadata_resourceTypesScope :: Lens' OrganizationCustomRuleMetadata (Maybe [Text])

The type of the Amazon Web Services resource that was evaluated.

organizationCustomRuleMetadata_tagValueScope :: Lens' OrganizationCustomRuleMetadata (Maybe Text)

The optional part of a key-value pair that make up a tag. A value acts as a descriptor within a tag category (key).

organizationCustomRuleMetadata_lambdaFunctionArn :: Lens' OrganizationCustomRuleMetadata Text

The lambda function ARN.

organizationCustomRuleMetadata_organizationConfigRuleTriggerTypes :: Lens' OrganizationCustomRuleMetadata [OrganizationConfigRuleTriggerType]

The type of notification that triggers Config to run an evaluation for a rule. You can specify the following notification types:

ConfigurationItemChangeNotification - Triggers an evaluation when Config delivers a configuration item as a result of a resource change.

OversizedConfigurationItemChangeNotification - Triggers an evaluation when Config delivers an oversized configuration item. Config may generate this notification type when a resource changes and the notification exceeds the maximum size allowed by Amazon SNS.

ScheduledNotification - Triggers a periodic evaluation at the frequency specified for MaximumExecutionFrequency.