Network.AWS.Config.Types.OrganizationManagedRuleMetadata

Description

Documentation

data OrganizationManagedRuleMetadata

An object that specifies organization managed rule metadata such as resource type and ID of Amazon Web Services resource along with the rule identifier. It also provides the frequency with which you want Config to run evaluations for the rule if the trigger type is periodic.

See: newOrganizationManagedRuleMetadata smart constructor.

Constructors

OrganizationManagedRuleMetadata'

Fields

tagKeyScope :: Maybe Text

One part of a key-value pair that make up a tag. A key is a general label that acts like a category for more specific tag values.

maximumExecutionFrequency :: Maybe MaximumExecutionFrequency

The maximum frequency with which Config runs evaluations for a rule. You are using an Config managed rule that is triggered at a periodic frequency.

By default, rules with a periodic trigger are evaluated every 24 hours. To change the frequency, specify a valid value for the MaximumExecutionFrequency parameter.

resourceIdScope :: Maybe Text

The ID of the Amazon Web Services resource that was evaluated.

inputParameters :: Maybe Text

A string, in JSON format, that is passed to organization config rule Lambda function.

description :: Maybe Text

The description that you provide for organization config rule.

resourceTypesScope :: Maybe [Text]

The type of the Amazon Web Services resource that was evaluated.

tagValueScope :: Maybe Text

The optional part of a key-value pair that make up a tag. A value acts as a descriptor within a tag category (key).

ruleIdentifier :: Text

For organization config managed rules, a predefined identifier from a list. For example, IAM_PASSWORD_POLICY is a managed rule. To reference a managed rule, see Using Config managed rules.

newOrganizationManagedRuleMetadata

Arguments

:: Text
$sel:ruleIdentifier:OrganizationManagedRuleMetadata'

-> OrganizationManagedRuleMetadata

Create a value of OrganizationManagedRuleMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tagKeyScope:OrganizationManagedRuleMetadata', organizationManagedRuleMetadata_tagKeyScope - One part of a key-value pair that make up a tag. A key is a general label that acts like a category for more specific tag values.

$sel:maximumExecutionFrequency:OrganizationManagedRuleMetadata', organizationManagedRuleMetadata_maximumExecutionFrequency - The maximum frequency with which Config runs evaluations for a rule. You are using an Config managed rule that is triggered at a periodic frequency.

By default, rules with a periodic trigger are evaluated every 24 hours. To change the frequency, specify a valid value for the MaximumExecutionFrequency parameter.

$sel:resourceIdScope:OrganizationManagedRuleMetadata', organizationManagedRuleMetadata_resourceIdScope - The ID of the Amazon Web Services resource that was evaluated.

$sel:inputParameters:OrganizationManagedRuleMetadata', organizationManagedRuleMetadata_inputParameters - A string, in JSON format, that is passed to organization config rule Lambda function.

$sel:description:OrganizationManagedRuleMetadata', organizationManagedRuleMetadata_description - The description that you provide for organization config rule.

$sel:resourceTypesScope:OrganizationManagedRuleMetadata', organizationManagedRuleMetadata_resourceTypesScope - The type of the Amazon Web Services resource that was evaluated.

$sel:tagValueScope:OrganizationManagedRuleMetadata', organizationManagedRuleMetadata_tagValueScope - The optional part of a key-value pair that make up a tag. A value acts as a descriptor within a tag category (key).

$sel:ruleIdentifier:OrganizationManagedRuleMetadata', organizationManagedRuleMetadata_ruleIdentifier - For organization config managed rules, a predefined identifier from a list. For example, IAM_PASSWORD_POLICY is a managed rule. To reference a managed rule, see Using Config managed rules.

organizationManagedRuleMetadata_tagKeyScope :: Lens' OrganizationManagedRuleMetadata (Maybe Text)

One part of a key-value pair that make up a tag. A key is a general label that acts like a category for more specific tag values.

organizationManagedRuleMetadata_maximumExecutionFrequency :: Lens' OrganizationManagedRuleMetadata (Maybe MaximumExecutionFrequency)

The maximum frequency with which Config runs evaluations for a rule. You are using an Config managed rule that is triggered at a periodic frequency.

By default, rules with a periodic trigger are evaluated every 24 hours. To change the frequency, specify a valid value for the MaximumExecutionFrequency parameter.

organizationManagedRuleMetadata_resourceIdScope :: Lens' OrganizationManagedRuleMetadata (Maybe Text)

The ID of the Amazon Web Services resource that was evaluated.

organizationManagedRuleMetadata_inputParameters :: Lens' OrganizationManagedRuleMetadata (Maybe Text)

A string, in JSON format, that is passed to organization config rule Lambda function.

organizationManagedRuleMetadata_description :: Lens' OrganizationManagedRuleMetadata (Maybe Text)

The description that you provide for organization config rule.

organizationManagedRuleMetadata_resourceTypesScope :: Lens' OrganizationManagedRuleMetadata (Maybe [Text])

The type of the Amazon Web Services resource that was evaluated.

organizationManagedRuleMetadata_tagValueScope :: Lens' OrganizationManagedRuleMetadata (Maybe Text)

The optional part of a key-value pair that make up a tag. A value acts as a descriptor within a tag category (key).

organizationManagedRuleMetadata_ruleIdentifier :: Lens' OrganizationManagedRuleMetadata Text

For organization config managed rules, a predefined identifier from a list. For example, IAM_PASSWORD_POLICY is a managed rule. To reference a managed rule, see Using Config managed rules.