Network.AWS.Config.Types.PendingAggregationRequest

Description

Documentation

data PendingAggregationRequest

An object that represents the account ID and region of an aggregator account that is requesting authorization but is not yet authorized.

See: newPendingAggregationRequest smart constructor.

Constructors

PendingAggregationRequest'

Fields

requesterAwsRegion :: Maybe Text

The region requesting to aggregate data.

requesterAccountId :: Maybe Text

The 12-digit account ID of the account requesting to aggregate data.

newPendingAggregationRequest :: PendingAggregationRequest

Create a value of PendingAggregationRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requesterAwsRegion:PendingAggregationRequest', pendingAggregationRequest_requesterAwsRegion - The region requesting to aggregate data.

$sel:requesterAccountId:PendingAggregationRequest', pendingAggregationRequest_requesterAccountId - The 12-digit account ID of the account requesting to aggregate data.

pendingAggregationRequest_requesterAwsRegion :: Lens' PendingAggregationRequest (Maybe Text)

The region requesting to aggregate data.

pendingAggregationRequest_requesterAccountId :: Lens' PendingAggregationRequest (Maybe Text)

The 12-digit account ID of the account requesting to aggregate data.