Network.AWS.Config.Types.RecordingGroup

Description

Documentation

data RecordingGroup

Specifies the types of Amazon Web Services resource for which Config records configuration changes.

In the recording group, you specify whether all supported types or specific types of resources are recorded.

By default, Config records configuration changes for all supported types of regional resources that Config discovers in the region in which it is running. Regional resources are tied to a region and can be used only in that region. Examples of regional resources are EC2 instances and EBS volumes.

You can also have Config record configuration changes for supported types of global resources (for example, IAM resources). Global resources are not tied to an individual region and can be used in all regions.

The configuration details for any global resource are the same in all regions. If you customize Config in multiple regions to record global resources, it will create multiple configuration items each time a global resource changes: one configuration item for each region. These configuration items will contain identical data. To prevent duplicate configuration items, you should consider customizing Config in only one region to record global resources, unless you want the configuration items to be available in multiple regions.

If you don't want Config to record all resources, you can specify which types of resources it will record with the resourceTypes parameter.

For a list of supported resource types, see Supported Resource Types.

For more information, see Selecting Which Resources Config Records.

See: newRecordingGroup smart constructor.

Constructors

RecordingGroup'

Fields

allSupported :: Maybe Bool

Specifies whether Config records configuration changes for every supported type of regional resource.

If you set this option to true, when Config adds support for a new type of regional resource, it starts recording resources of that type automatically.

If you set this option to true, you cannot enumerate a list of resourceTypes.

resourceTypes :: Maybe [ResourceType]

A comma-separated list that specifies the types of Amazon Web Services resources for which Config records configuration changes (for example, AWS::EC2::Instance or AWS::CloudTrail::Trail).

To record all configuration changes, you must set the allSupported option to true.

If you set this option to false, when Config adds support for a new type of resource, it will not record resources of that type unless you manually add that type to your recording group.

For a list of valid resourceTypes values, see the __resourceType Value__ column in Supported Amazon Web Services resource Types.

includeGlobalResourceTypes :: Maybe Bool

Specifies whether Config includes all supported types of global resources (for example, IAM resources) with the resources that it records.

Before you can set this option to true, you must set the allSupported option to true.

If you set this option to true, when Config adds support for a new type of global resource, it starts recording resources of that type automatically.

The configuration details for any global resource are the same in all regions. To prevent duplicate configuration items, you should consider customizing Config in only one region to record global resources.

newRecordingGroup :: RecordingGroup

Create a value of RecordingGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:allSupported:RecordingGroup', recordingGroup_allSupported - Specifies whether Config records configuration changes for every supported type of regional resource.

If you set this option to true, when Config adds support for a new type of regional resource, it starts recording resources of that type automatically.

If you set this option to true, you cannot enumerate a list of resourceTypes.

$sel:resourceTypes:RecordingGroup', recordingGroup_resourceTypes - A comma-separated list that specifies the types of Amazon Web Services resources for which Config records configuration changes (for example, AWS::EC2::Instance or AWS::CloudTrail::Trail).

To record all configuration changes, you must set the allSupported option to true.

If you set this option to false, when Config adds support for a new type of resource, it will not record resources of that type unless you manually add that type to your recording group.

For a list of valid resourceTypes values, see the __resourceType Value__ column in Supported Amazon Web Services resource Types.

$sel:includeGlobalResourceTypes:RecordingGroup', recordingGroup_includeGlobalResourceTypes - Specifies whether Config includes all supported types of global resources (for example, IAM resources) with the resources that it records.

Before you can set this option to true, you must set the allSupported option to true.

If you set this option to true, when Config adds support for a new type of global resource, it starts recording resources of that type automatically.

The configuration details for any global resource are the same in all regions. To prevent duplicate configuration items, you should consider customizing Config in only one region to record global resources.

recordingGroup_allSupported :: Lens' RecordingGroup (Maybe Bool)

Specifies whether Config records configuration changes for every supported type of regional resource.

If you set this option to true, when Config adds support for a new type of regional resource, it starts recording resources of that type automatically.

If you set this option to true, you cannot enumerate a list of resourceTypes.

recordingGroup_resourceTypes :: Lens' RecordingGroup (Maybe [ResourceType])

A comma-separated list that specifies the types of Amazon Web Services resources for which Config records configuration changes (for example, AWS::EC2::Instance or AWS::CloudTrail::Trail).

To record all configuration changes, you must set the allSupported option to true.

If you set this option to false, when Config adds support for a new type of resource, it will not record resources of that type unless you manually add that type to your recording group.

For a list of valid resourceTypes values, see the __resourceType Value__ column in Supported Amazon Web Services resource Types.

recordingGroup_includeGlobalResourceTypes :: Lens' RecordingGroup (Maybe Bool)

Specifies whether Config includes all supported types of global resources (for example, IAM resources) with the resources that it records.

Before you can set this option to true, you must set the allSupported option to true.

If you set this option to true, when Config adds support for a new type of global resource, it starts recording resources of that type automatically.

The configuration details for any global resource are the same in all regions. To prevent duplicate configuration items, you should consider customizing Config in only one region to record global resources.