Network.AWS.Config.Types.Relationship

Description

Documentation

data Relationship

The relationship of the related resource to the main resource.

See: newRelationship smart constructor.

Constructors

Relationship'

Fields

resourceId :: Maybe Text

The ID of the related resource (for example, sg-xxxxxx).

resourceType :: Maybe ResourceType

The resource type of the related resource.

relationshipName :: Maybe Text

The type of relationship with the related resource.

resourceName :: Maybe Text

The custom name of the related resource, if available.

newRelationship :: Relationship

Create a value of Relationship with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceId:Relationship', relationship_resourceId - The ID of the related resource (for example, sg-xxxxxx).

$sel:resourceType:Relationship', relationship_resourceType - The resource type of the related resource.

$sel:relationshipName:Relationship', relationship_relationshipName - The type of relationship with the related resource.

$sel:resourceName:Relationship', relationship_resourceName - The custom name of the related resource, if available.

relationship_resourceId :: Lens' Relationship (Maybe Text)

The ID of the related resource (for example, sg-xxxxxx).

relationship_resourceType :: Lens' Relationship (Maybe ResourceType)

The resource type of the related resource.

relationship_relationshipName :: Lens' Relationship (Maybe Text)

The type of relationship with the related resource.

relationship_resourceName :: Lens' Relationship (Maybe Text)

The custom name of the related resource, if available.