Network.AWS.Config.Types.RemediationConfiguration

Description

Documentation

data RemediationConfiguration

An object that represents the details about the remediation configuration that includes the remediation action, parameters, and data to execute the action.

See: newRemediationConfiguration smart constructor.

Constructors

RemediationConfiguration'

Fields

executionControls :: Maybe ExecutionControls

An ExecutionControls object.

retryAttemptSeconds :: Maybe Natural

Maximum time in seconds that Config runs auto-remediation. If you do not select a number, the default is 60 seconds.

For example, if you specify RetryAttemptSeconds as 50 seconds and MaximumAutomaticAttempts as 5, Config will run auto-remediations 5 times within 50 seconds before throwing an exception.

targetVersion :: Maybe Text

Version of the target. For example, version of the SSM document.

If you make backward incompatible changes to the SSM document, you must call PutRemediationConfiguration API again to ensure the remediations can run.

arn :: Maybe Text

Amazon Resource Name (ARN) of remediation configuration.

automatic :: Maybe Bool

The remediation is triggered automatically.

resourceType :: Maybe Text

The type of a resource.

createdByService :: Maybe Text

Name of the service that owns the service linked rule, if applicable.

maximumAutomaticAttempts :: Maybe Natural

The maximum number of failed attempts for auto-remediation. If you do not select a number, the default is 5.

For example, if you specify MaximumAutomaticAttempts as 5 with RetryAttemptSeconds as 50 seconds, Config will put a RemediationException on your behalf for the failing resource after the 5th failed attempt within 50 seconds.

parameters :: Maybe (HashMap Text RemediationParameterValue)

An object of the RemediationParameterValue.

configRuleName :: Text

The name of the Config rule.

targetType :: RemediationTargetType

The type of the target. Target executes remediation. For example, SSM document.

targetId :: Text

Target ID is the name of the public document.

newRemediationConfiguration

Arguments

:: Text
$sel:configRuleName:RemediationConfiguration'

-> RemediationTargetType
$sel:targetType:RemediationConfiguration'

-> Text
$sel:targetId:RemediationConfiguration'

-> RemediationConfiguration

Create a value of RemediationConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:executionControls:RemediationConfiguration', remediationConfiguration_executionControls - An ExecutionControls object.

$sel:retryAttemptSeconds:RemediationConfiguration', remediationConfiguration_retryAttemptSeconds - Maximum time in seconds that Config runs auto-remediation. If you do not select a number, the default is 60 seconds.

For example, if you specify RetryAttemptSeconds as 50 seconds and MaximumAutomaticAttempts as 5, Config will run auto-remediations 5 times within 50 seconds before throwing an exception.

$sel:targetVersion:RemediationConfiguration', remediationConfiguration_targetVersion - Version of the target. For example, version of the SSM document.

If you make backward incompatible changes to the SSM document, you must call PutRemediationConfiguration API again to ensure the remediations can run.

$sel:arn:RemediationConfiguration', remediationConfiguration_arn - Amazon Resource Name (ARN) of remediation configuration.

$sel:automatic:RemediationConfiguration', remediationConfiguration_automatic - The remediation is triggered automatically.

$sel:resourceType:RemediationConfiguration', remediationConfiguration_resourceType - The type of a resource.

$sel:createdByService:RemediationConfiguration', remediationConfiguration_createdByService - Name of the service that owns the service linked rule, if applicable.

$sel:maximumAutomaticAttempts:RemediationConfiguration', remediationConfiguration_maximumAutomaticAttempts - The maximum number of failed attempts for auto-remediation. If you do not select a number, the default is 5.

For example, if you specify MaximumAutomaticAttempts as 5 with RetryAttemptSeconds as 50 seconds, Config will put a RemediationException on your behalf for the failing resource after the 5th failed attempt within 50 seconds.

$sel:parameters:RemediationConfiguration', remediationConfiguration_parameters - An object of the RemediationParameterValue.

$sel:configRuleName:RemediationConfiguration', remediationConfiguration_configRuleName - The name of the Config rule.

$sel:targetType:RemediationConfiguration', remediationConfiguration_targetType - The type of the target. Target executes remediation. For example, SSM document.

$sel:targetId:RemediationConfiguration', remediationConfiguration_targetId - Target ID is the name of the public document.

remediationConfiguration_executionControls :: Lens' RemediationConfiguration (Maybe ExecutionControls)

An ExecutionControls object.

remediationConfiguration_retryAttemptSeconds :: Lens' RemediationConfiguration (Maybe Natural)

Maximum time in seconds that Config runs auto-remediation. If you do not select a number, the default is 60 seconds.

For example, if you specify RetryAttemptSeconds as 50 seconds and MaximumAutomaticAttempts as 5, Config will run auto-remediations 5 times within 50 seconds before throwing an exception.

remediationConfiguration_targetVersion :: Lens' RemediationConfiguration (Maybe Text)

Version of the target. For example, version of the SSM document.

If you make backward incompatible changes to the SSM document, you must call PutRemediationConfiguration API again to ensure the remediations can run.

remediationConfiguration_arn :: Lens' RemediationConfiguration (Maybe Text)

Amazon Resource Name (ARN) of remediation configuration.

remediationConfiguration_automatic :: Lens' RemediationConfiguration (Maybe Bool)

The remediation is triggered automatically.

remediationConfiguration_resourceType :: Lens' RemediationConfiguration (Maybe Text)

The type of a resource.

remediationConfiguration_createdByService :: Lens' RemediationConfiguration (Maybe Text)

Name of the service that owns the service linked rule, if applicable.

remediationConfiguration_maximumAutomaticAttempts :: Lens' RemediationConfiguration (Maybe Natural)

The maximum number of failed attempts for auto-remediation. If you do not select a number, the default is 5.

For example, if you specify MaximumAutomaticAttempts as 5 with RetryAttemptSeconds as 50 seconds, Config will put a RemediationException on your behalf for the failing resource after the 5th failed attempt within 50 seconds.

remediationConfiguration_parameters :: Lens' RemediationConfiguration (Maybe (HashMap Text RemediationParameterValue))

An object of the RemediationParameterValue.

remediationConfiguration_configRuleName :: Lens' RemediationConfiguration Text

The name of the Config rule.

remediationConfiguration_targetType :: Lens' RemediationConfiguration RemediationTargetType

The type of the target. Target executes remediation. For example, SSM document.

remediationConfiguration_targetId :: Lens' RemediationConfiguration Text

Target ID is the name of the public document.