Network.AWS.Config.Types.RetentionConfiguration

Description

Documentation

data RetentionConfiguration

An object with the name of the retention configuration and the retention period in days. The object stores the configuration for data retention in Config.

See: newRetentionConfiguration smart constructor.

Constructors

RetentionConfiguration'

Fields

name :: Text

The name of the retention configuration object.

retentionPeriodInDays :: Natural

Number of days Config stores your historical information.

Currently, only applicable to the configuration item history.

newRetentionConfiguration

Arguments

:: Text
$sel:name:RetentionConfiguration'

-> Natural
$sel:retentionPeriodInDays:RetentionConfiguration'

-> RetentionConfiguration

Create a value of RetentionConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:RetentionConfiguration', retentionConfiguration_name - The name of the retention configuration object.

$sel:retentionPeriodInDays:RetentionConfiguration', retentionConfiguration_retentionPeriodInDays - Number of days Config stores your historical information.

Currently, only applicable to the configuration item history.

retentionConfiguration_name :: Lens' RetentionConfiguration Text

The name of the retention configuration object.

retentionConfiguration_retentionPeriodInDays :: Lens' RetentionConfiguration Natural

Number of days Config stores your historical information.

Currently, only applicable to the configuration item history.