Network.AWS.Config.Types.Source

Description

Documentation

data Source

Provides the Config rule owner (Amazon Web Services or customer), the rule identifier, and the events that trigger the evaluation of your Amazon Web Services resources.

See: newSource smart constructor.

Constructors

Source'

Fields

sourceDetails :: Maybe [SourceDetail]

Provides the source and type of the event that causes Config to evaluate your Amazon Web Services resources.

owner :: Owner

Indicates whether Amazon Web Services or the customer owns and manages the Config rule.

sourceIdentifier :: Text

For Config managed rules, a predefined identifier from a list. For example, IAM_PASSWORD_POLICY is a managed rule. To reference a managed rule, see Using Config managed rules.

For custom rules, the identifier is the Amazon Resource Name (ARN) of the rule's Lambda function, such as arn:aws:lambda:us-east-2:123456789012:function:custom_rule_name.

newSource

Arguments

:: Owner
$sel:owner:Source'

-> Text
$sel:sourceIdentifier:Source'

-> Source

Create a value of Source with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceDetails:Source', source_sourceDetails - Provides the source and type of the event that causes Config to evaluate your Amazon Web Services resources.

$sel:owner:Source', source_owner - Indicates whether Amazon Web Services or the customer owns and manages the Config rule.

$sel:sourceIdentifier:Source', source_sourceIdentifier - For Config managed rules, a predefined identifier from a list. For example, IAM_PASSWORD_POLICY is a managed rule. To reference a managed rule, see Using Config managed rules.

For custom rules, the identifier is the Amazon Resource Name (ARN) of the rule's Lambda function, such as arn:aws:lambda:us-east-2:123456789012:function:custom_rule_name.

source_sourceDetails :: Lens' Source (Maybe [SourceDetail])

Provides the source and type of the event that causes Config to evaluate your Amazon Web Services resources.

source_owner :: Lens' Source Owner

Indicates whether Amazon Web Services or the customer owns and manages the Config rule.

source_sourceIdentifier :: Lens' Source Text

For Config managed rules, a predefined identifier from a list. For example, IAM_PASSWORD_POLICY is a managed rule. To reference a managed rule, see Using Config managed rules.

For custom rules, the identifier is the Amazon Resource Name (ARN) of the rule's Lambda function, such as arn:aws:lambda:us-east-2:123456789012:function:custom_rule_name.