Network.AWS.Config.Types.SsmControls

Description

Documentation

data SsmControls

Amazon Web Services Systems Manager (SSM) specific remediation controls.

See: newSsmControls smart constructor.

Constructors

SsmControls'

Fields

errorPercentage :: Maybe Natural

The percentage of errors that are allowed before SSM stops running automations on non-compliant resources for that specific rule. You can specify a percentage of errors, for example 10%. If you do not specifiy a percentage, the default is 50%. For example, if you set the ErrorPercentage to 40% for 10 non-compliant resources, then SSM stops running the automations when the fifth error is received.

concurrentExecutionRatePercentage :: Maybe Natural

The maximum percentage of remediation actions allowed to run in parallel on the non-compliant resources for that specific rule. You can specify a percentage, such as 10%. The default value is 10.

newSsmControls :: SsmControls

Create a value of SsmControls with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:errorPercentage:SsmControls', ssmControls_errorPercentage - The percentage of errors that are allowed before SSM stops running automations on non-compliant resources for that specific rule. You can specify a percentage of errors, for example 10%. If you do not specifiy a percentage, the default is 50%. For example, if you set the ErrorPercentage to 40% for 10 non-compliant resources, then SSM stops running the automations when the fifth error is received.

$sel:concurrentExecutionRatePercentage:SsmControls', ssmControls_concurrentExecutionRatePercentage - The maximum percentage of remediation actions allowed to run in parallel on the non-compliant resources for that specific rule. You can specify a percentage, such as 10%. The default value is 10.

ssmControls_errorPercentage :: Lens' SsmControls (Maybe Natural)

The percentage of errors that are allowed before SSM stops running automations on non-compliant resources for that specific rule. You can specify a percentage of errors, for example 10%. If you do not specifiy a percentage, the default is 50%. For example, if you set the ErrorPercentage to 40% for 10 non-compliant resources, then SSM stops running the automations when the fifth error is received.

ssmControls_concurrentExecutionRatePercentage :: Lens' SsmControls (Maybe Natural)

The maximum percentage of remediation actions allowed to run in parallel on the non-compliant resources for that specific rule. You can specify a percentage, such as 10%. The default value is 10.