Network.AWS.Config.Types.StatusDetailFilters

Description

Documentation

data StatusDetailFilters

Status filter object to filter results based on specific member account ID or status type for an organization config rule.

See: newStatusDetailFilters smart constructor.

Constructors

StatusDetailFilters'

Fields

accountId :: Maybe Text

The 12-digit account ID of the member account within an organization.

memberAccountRuleStatus :: Maybe MemberAccountRuleStatus

Indicates deployment status for config rule in the member account. When master account calls PutOrganizationConfigRule action for the first time, config rule status is created in the member account. When master account calls PutOrganizationConfigRule action for the second time, config rule status is updated in the member account. Config rule status is deleted when the master account deletes OrganizationConfigRule and disables service access for config-multiaccountsetup.amazonaws.com.

Config sets the state of the rule to:

CREATE_SUCCESSFUL when config rule has been created in the member account.

CREATE_IN_PROGRESS when config rule is being created in the member account.

CREATE_FAILED when config rule creation has failed in the member account.

DELETE_FAILED when config rule deletion has failed in the member account.

DELETE_IN_PROGRESS when config rule is being deleted in the member account.

DELETE_SUCCESSFUL when config rule has been deleted in the member account.

UPDATE_SUCCESSFUL when config rule has been updated in the member account.

UPDATE_IN_PROGRESS when config rule is being updated in the member account.

UPDATE_FAILED when config rule deletion has failed in the member account.

newStatusDetailFilters :: StatusDetailFilters

Create a value of StatusDetailFilters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountId:StatusDetailFilters', statusDetailFilters_accountId - The 12-digit account ID of the member account within an organization.

$sel:memberAccountRuleStatus:StatusDetailFilters', statusDetailFilters_memberAccountRuleStatus - Indicates deployment status for config rule in the member account. When master account calls PutOrganizationConfigRule action for the first time, config rule status is created in the member account. When master account calls PutOrganizationConfigRule action for the second time, config rule status is updated in the member account. Config rule status is deleted when the master account deletes OrganizationConfigRule and disables service access for config-multiaccountsetup.amazonaws.com.

Config sets the state of the rule to:

CREATE_SUCCESSFUL when config rule has been created in the member account.

CREATE_IN_PROGRESS when config rule is being created in the member account.

CREATE_FAILED when config rule creation has failed in the member account.

DELETE_FAILED when config rule deletion has failed in the member account.

DELETE_IN_PROGRESS when config rule is being deleted in the member account.

DELETE_SUCCESSFUL when config rule has been deleted in the member account.

UPDATE_SUCCESSFUL when config rule has been updated in the member account.

UPDATE_IN_PROGRESS when config rule is being updated in the member account.

UPDATE_FAILED when config rule deletion has failed in the member account.

statusDetailFilters_accountId :: Lens' StatusDetailFilters (Maybe Text)

The 12-digit account ID of the member account within an organization.

statusDetailFilters_memberAccountRuleStatus :: Lens' StatusDetailFilters (Maybe MemberAccountRuleStatus)

Indicates deployment status for config rule in the member account. When master account calls PutOrganizationConfigRule action for the first time, config rule status is created in the member account. When master account calls PutOrganizationConfigRule action for the second time, config rule status is updated in the member account. Config rule status is deleted when the master account deletes OrganizationConfigRule and disables service access for config-multiaccountsetup.amazonaws.com.

Config sets the state of the rule to:

CREATE_SUCCESSFUL when config rule has been created in the member account.

CREATE_IN_PROGRESS when config rule is being created in the member account.

CREATE_FAILED when config rule creation has failed in the member account.

DELETE_FAILED when config rule deletion has failed in the member account.

DELETE_IN_PROGRESS when config rule is being deleted in the member account.

DELETE_SUCCESSFUL when config rule has been deleted in the member account.

UPDATE_SUCCESSFUL when config rule has been updated in the member account.

UPDATE_IN_PROGRESS when config rule is being updated in the member account.

UPDATE_FAILED when config rule deletion has failed in the member account.