Network.AWS.Connect.AssociateRoutingProfileQueues

Description

Associates a set of queues with a routing profile.

Creating a Request

data AssociateRoutingProfileQueues

See: newAssociateRoutingProfileQueues smart constructor.

Constructors

AssociateRoutingProfileQueues'

Fields

instanceId :: Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

routingProfileId :: Text

The identifier of the routing profile.

queueConfigs :: NonEmpty RoutingProfileQueueConfig

The queues to associate with this routing profile.

newAssociateRoutingProfileQueues

Arguments

:: Text
$sel:instanceId:AssociateRoutingProfileQueues'

-> Text
$sel:routingProfileId:AssociateRoutingProfileQueues'

-> NonEmpty RoutingProfileQueueConfig
$sel:queueConfigs:AssociateRoutingProfileQueues'

-> AssociateRoutingProfileQueues

Create a value of AssociateRoutingProfileQueues with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceId:AssociateRoutingProfileQueues', associateRoutingProfileQueues_instanceId - The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

$sel:routingProfileId:AssociateRoutingProfileQueues', associateRoutingProfileQueues_routingProfileId - The identifier of the routing profile.

$sel:queueConfigs:AssociateRoutingProfileQueues', associateRoutingProfileQueues_queueConfigs - The queues to associate with this routing profile.

Request Lenses

associateRoutingProfileQueues_instanceId :: Lens' AssociateRoutingProfileQueues Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

associateRoutingProfileQueues_routingProfileId :: Lens' AssociateRoutingProfileQueues Text

The identifier of the routing profile.

associateRoutingProfileQueues_queueConfigs :: Lens' AssociateRoutingProfileQueues (NonEmpty RoutingProfileQueueConfig)

The queues to associate with this routing profile.

Destructuring the Response

data AssociateRoutingProfileQueuesResponse

See: newAssociateRoutingProfileQueuesResponse smart constructor.

Constructors

AssociateRoutingProfileQueuesResponse'

newAssociateRoutingProfileQueuesResponse :: AssociateRoutingProfileQueuesResponse

Create a value of AssociateRoutingProfileQueuesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.