Network.AWS.Connect.CreateContactFlow

Description

Creates a contact flow for the specified Amazon Connect instance.

You can also create and update contact flows using the Amazon Connect Flow language.

Creating a Request

data CreateContactFlow

See: newCreateContactFlow smart constructor.

Constructors

CreateContactFlow'

Fields

description :: Maybe Text

The description of the contact flow.

tags :: Maybe (HashMap Text Text)

One or more tags.

instanceId :: Text

The identifier of the Amazon Connect instance.

name :: Text

The name of the contact flow.

type' :: ContactFlowType

The type of the contact flow. For descriptions of the available types, see Choose a Contact Flow Type in the Amazon Connect Administrator Guide.

content :: Text

The content of the contact flow.

newCreateContactFlow

Arguments

:: Text
$sel:instanceId:CreateContactFlow'

-> Text
$sel:name:CreateContactFlow'

-> ContactFlowType
$sel:type':CreateContactFlow'

-> Text
$sel:content:CreateContactFlow'

-> CreateContactFlow

Create a value of CreateContactFlow with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:CreateContactFlow', createContactFlow_description - The description of the contact flow.

$sel:tags:CreateContactFlow', createContactFlow_tags - One or more tags.

$sel:instanceId:CreateContactFlow', createContactFlow_instanceId - The identifier of the Amazon Connect instance.

$sel:name:CreateContactFlow', createContactFlow_name - The name of the contact flow.

$sel:type':CreateContactFlow', createContactFlow_type - The type of the contact flow. For descriptions of the available types, see Choose a Contact Flow Type in the Amazon Connect Administrator Guide.

$sel:content:CreateContactFlow', createContactFlow_content - The content of the contact flow.

Request Lenses

createContactFlow_description :: Lens' CreateContactFlow (Maybe Text)

The description of the contact flow.

createContactFlow_tags :: Lens' CreateContactFlow (Maybe (HashMap Text Text))

One or more tags.

createContactFlow_instanceId :: Lens' CreateContactFlow Text

The identifier of the Amazon Connect instance.

createContactFlow_name :: Lens' CreateContactFlow Text

The name of the contact flow.

createContactFlow_type :: Lens' CreateContactFlow ContactFlowType

The type of the contact flow. For descriptions of the available types, see Choose a Contact Flow Type in the Amazon Connect Administrator Guide.

createContactFlow_content :: Lens' CreateContactFlow Text

The content of the contact flow.

Destructuring the Response

data CreateContactFlowResponse

See: newCreateContactFlowResponse smart constructor.

Constructors

CreateContactFlowResponse'

Fields

contactFlowArn :: Maybe Text

The Amazon Resource Name (ARN) of the contact flow.

contactFlowId :: Maybe Text

The identifier of the contact flow.

httpStatus :: Int

The response's http status code.

newCreateContactFlowResponse

Arguments

:: Int
$sel:httpStatus:CreateContactFlowResponse'

-> CreateContactFlowResponse

Create a value of CreateContactFlowResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:contactFlowArn:CreateContactFlowResponse', createContactFlowResponse_contactFlowArn - The Amazon Resource Name (ARN) of the contact flow.

$sel:contactFlowId:CreateContactFlowResponse', createContactFlowResponse_contactFlowId - The identifier of the contact flow.

$sel:httpStatus:CreateContactFlowResponse', createContactFlowResponse_httpStatus - The response's http status code.

Response Lenses

createContactFlowResponse_contactFlowArn :: Lens' CreateContactFlowResponse (Maybe Text)

The Amazon Resource Name (ARN) of the contact flow.

createContactFlowResponse_contactFlowId :: Lens' CreateContactFlowResponse (Maybe Text)

The identifier of the contact flow.

createContactFlowResponse_httpStatus :: Lens' CreateContactFlowResponse Int

The response's http status code.