Network.AWS.Connect.CreateIntegrationAssociation

Description

Creates an AWS resource association with an Amazon Connect instance.

Creating a Request

data CreateIntegrationAssociation

See: newCreateIntegrationAssociation smart constructor.

Constructors

CreateIntegrationAssociation'

Fields

sourceApplicationName :: Maybe Text

The name of the external application. This field is only required for the EVENT integration type.

sourceApplicationUrl :: Maybe Text

The URL for the external application. This field is only required for the EVENT integration type.

tags :: Maybe (HashMap Text Text)

One or more tags.

sourceType :: Maybe SourceType

The type of the data source. This field is only required for the EVENT integration type.

instanceId :: Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

integrationType :: IntegrationType

The type of information to be ingested.

integrationArn :: Text

The Amazon Resource Name (ARN) of the integration.

newCreateIntegrationAssociation

Arguments

:: Text
$sel:instanceId:CreateIntegrationAssociation'

-> IntegrationType
$sel:integrationType:CreateIntegrationAssociation'

-> Text
$sel:integrationArn:CreateIntegrationAssociation'

-> CreateIntegrationAssociation

Create a value of CreateIntegrationAssociation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceApplicationName:CreateIntegrationAssociation', createIntegrationAssociation_sourceApplicationName - The name of the external application. This field is only required for the EVENT integration type.

$sel:sourceApplicationUrl:CreateIntegrationAssociation', createIntegrationAssociation_sourceApplicationUrl - The URL for the external application. This field is only required for the EVENT integration type.

$sel:tags:CreateIntegrationAssociation', createIntegrationAssociation_tags - One or more tags.

$sel:sourceType:CreateIntegrationAssociation', createIntegrationAssociation_sourceType - The type of the data source. This field is only required for the EVENT integration type.

$sel:instanceId:CreateIntegrationAssociation', createIntegrationAssociation_instanceId - The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

$sel:integrationType:CreateIntegrationAssociation', createIntegrationAssociation_integrationType - The type of information to be ingested.

$sel:integrationArn:CreateIntegrationAssociation', createIntegrationAssociation_integrationArn - The Amazon Resource Name (ARN) of the integration.

Request Lenses

createIntegrationAssociation_sourceApplicationName :: Lens' CreateIntegrationAssociation (Maybe Text)

The name of the external application. This field is only required for the EVENT integration type.

createIntegrationAssociation_sourceApplicationUrl :: Lens' CreateIntegrationAssociation (Maybe Text)

The URL for the external application. This field is only required for the EVENT integration type.

createIntegrationAssociation_tags :: Lens' CreateIntegrationAssociation (Maybe (HashMap Text Text))

One or more tags.

createIntegrationAssociation_sourceType :: Lens' CreateIntegrationAssociation (Maybe SourceType)

The type of the data source. This field is only required for the EVENT integration type.

createIntegrationAssociation_instanceId :: Lens' CreateIntegrationAssociation Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

createIntegrationAssociation_integrationType :: Lens' CreateIntegrationAssociation IntegrationType

The type of information to be ingested.

createIntegrationAssociation_integrationArn :: Lens' CreateIntegrationAssociation Text

The Amazon Resource Name (ARN) of the integration.

Destructuring the Response

data CreateIntegrationAssociationResponse

See: newCreateIntegrationAssociationResponse smart constructor.

Constructors

CreateIntegrationAssociationResponse'

Fields

integrationAssociationArn :: Maybe Text

The Amazon Resource Name (ARN) for the association.

integrationAssociationId :: Maybe Text

The identifier for the integration association.

httpStatus :: Int

The response's http status code.

newCreateIntegrationAssociationResponse

Arguments

:: Int
$sel:httpStatus:CreateIntegrationAssociationResponse'

-> CreateIntegrationAssociationResponse

Create a value of CreateIntegrationAssociationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:integrationAssociationArn:CreateIntegrationAssociationResponse', createIntegrationAssociationResponse_integrationAssociationArn - The Amazon Resource Name (ARN) for the association.

$sel:integrationAssociationId:CreateIntegrationAssociationResponse', createIntegrationAssociationResponse_integrationAssociationId - The identifier for the integration association.

$sel:httpStatus:CreateIntegrationAssociationResponse', createIntegrationAssociationResponse_httpStatus - The response's http status code.

Response Lenses

createIntegrationAssociationResponse_integrationAssociationArn :: Lens' CreateIntegrationAssociationResponse (Maybe Text)

The Amazon Resource Name (ARN) for the association.

createIntegrationAssociationResponse_integrationAssociationId :: Lens' CreateIntegrationAssociationResponse (Maybe Text)

The identifier for the integration association.

createIntegrationAssociationResponse_httpStatus :: Lens' CreateIntegrationAssociationResponse Int

The response's http status code.