Network.AWS.Connect.CreateQuickConnect

Description

Creates a quick connect for the specified Amazon Connect instance.

Creating a Request

data CreateQuickConnect

See: newCreateQuickConnect smart constructor.

Constructors

CreateQuickConnect'

Fields

description :: Maybe Text

The description of the quick connect.

tags :: Maybe (HashMap Text Text)

One or more tags.

instanceId :: Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

name :: Text

The name of the quick connect.

quickConnectConfig :: QuickConnectConfig

Configuration settings for the quick connect.

newCreateQuickConnect

Arguments

:: Text
$sel:instanceId:CreateQuickConnect'

-> Text
$sel:name:CreateQuickConnect'

-> QuickConnectConfig
$sel:quickConnectConfig:CreateQuickConnect'

-> CreateQuickConnect

Create a value of CreateQuickConnect with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:CreateQuickConnect', createQuickConnect_description - The description of the quick connect.

$sel:tags:CreateQuickConnect', createQuickConnect_tags - One or more tags.

$sel:instanceId:CreateQuickConnect', createQuickConnect_instanceId - The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

$sel:name:CreateQuickConnect', createQuickConnect_name - The name of the quick connect.

$sel:quickConnectConfig:CreateQuickConnect', createQuickConnect_quickConnectConfig - Configuration settings for the quick connect.

Request Lenses

createQuickConnect_description :: Lens' CreateQuickConnect (Maybe Text)

The description of the quick connect.

createQuickConnect_tags :: Lens' CreateQuickConnect (Maybe (HashMap Text Text))

One or more tags.

createQuickConnect_instanceId :: Lens' CreateQuickConnect Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

createQuickConnect_name :: Lens' CreateQuickConnect Text

The name of the quick connect.

createQuickConnect_quickConnectConfig :: Lens' CreateQuickConnect QuickConnectConfig

Configuration settings for the quick connect.

Destructuring the Response

data CreateQuickConnectResponse

See: newCreateQuickConnectResponse smart constructor.

Constructors

CreateQuickConnectResponse'

Fields

quickConnectId :: Maybe Text

The identifier for the quick connect.

quickConnectARN :: Maybe Text

The Amazon Resource Name (ARN) for the quick connect.

httpStatus :: Int

The response's http status code.

newCreateQuickConnectResponse

Arguments

:: Int
$sel:httpStatus:CreateQuickConnectResponse'

-> CreateQuickConnectResponse

Create a value of CreateQuickConnectResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:quickConnectId:CreateQuickConnectResponse', createQuickConnectResponse_quickConnectId - The identifier for the quick connect.

$sel:quickConnectARN:CreateQuickConnectResponse', createQuickConnectResponse_quickConnectARN - The Amazon Resource Name (ARN) for the quick connect.

$sel:httpStatus:CreateQuickConnectResponse', createQuickConnectResponse_httpStatus - The response's http status code.

Response Lenses

createQuickConnectResponse_quickConnectId :: Lens' CreateQuickConnectResponse (Maybe Text)

The identifier for the quick connect.

createQuickConnectResponse_quickConnectARN :: Lens' CreateQuickConnectResponse (Maybe Text)

The Amazon Resource Name (ARN) for the quick connect.

createQuickConnectResponse_httpStatus :: Lens' CreateQuickConnectResponse Int

The response's http status code.