Network.AWS.Connect.CreateUseCase

Description

Creates a use case for an integration association.

Creating a Request

data CreateUseCase

See: newCreateUseCase smart constructor.

Constructors

CreateUseCase'

Fields

tags :: Maybe (HashMap Text Text)

One or more tags.

instanceId :: Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

integrationAssociationId :: Text

The identifier for the integration association.

useCaseType :: UseCaseType

The type of use case to associate to the integration association. Each integration association can have only one of each use case type.

newCreateUseCase

Arguments

:: Text
$sel:instanceId:CreateUseCase'

-> Text
$sel:integrationAssociationId:CreateUseCase'

-> UseCaseType
$sel:useCaseType:CreateUseCase'

-> CreateUseCase

Create a value of CreateUseCase with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateUseCase', createUseCase_tags - One or more tags.

$sel:instanceId:CreateUseCase', createUseCase_instanceId - The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

$sel:integrationAssociationId:CreateUseCase', createUseCase_integrationAssociationId - The identifier for the integration association.

$sel:useCaseType:CreateUseCase', createUseCase_useCaseType - The type of use case to associate to the integration association. Each integration association can have only one of each use case type.

Request Lenses

createUseCase_tags :: Lens' CreateUseCase (Maybe (HashMap Text Text))

One or more tags.

createUseCase_instanceId :: Lens' CreateUseCase Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

createUseCase_integrationAssociationId :: Lens' CreateUseCase Text

The identifier for the integration association.

createUseCase_useCaseType :: Lens' CreateUseCase UseCaseType

The type of use case to associate to the integration association. Each integration association can have only one of each use case type.

Destructuring the Response

data CreateUseCaseResponse

See: newCreateUseCaseResponse smart constructor.

Constructors

CreateUseCaseResponse'

Fields

useCaseArn :: Maybe Text

The Amazon Resource Name (ARN) for the use case.

useCaseId :: Maybe Text

The identifier of the use case.

httpStatus :: Int

The response's http status code.

newCreateUseCaseResponse

Arguments

:: Int
$sel:httpStatus:CreateUseCaseResponse'

-> CreateUseCaseResponse

Create a value of CreateUseCaseResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:useCaseArn:CreateUseCaseResponse', createUseCaseResponse_useCaseArn - The Amazon Resource Name (ARN) for the use case.

$sel:useCaseId:CreateUseCaseResponse', createUseCaseResponse_useCaseId - The identifier of the use case.

$sel:httpStatus:CreateUseCaseResponse', createUseCaseResponse_httpStatus - The response's http status code.

Response Lenses

createUseCaseResponse_useCaseArn :: Lens' CreateUseCaseResponse (Maybe Text)

The Amazon Resource Name (ARN) for the use case.

createUseCaseResponse_useCaseId :: Lens' CreateUseCaseResponse (Maybe Text)

The identifier of the use case.

createUseCaseResponse_httpStatus :: Lens' CreateUseCaseResponse Int

The response's http status code.