Network.AWS.Connect.CreateUser

Description

Creates a user account for the specified Amazon Connect instance.

For information about how to create user accounts using the Amazon Connect console, see Add Users in the Amazon Connect Administrator Guide.

Creating a Request

data CreateUser

See: newCreateUser smart constructor.

Constructors

CreateUser'

Fields

identityInfo :: Maybe UserIdentityInfo

The information about the identity of the user.

hierarchyGroupId :: Maybe Text

The identifier of the hierarchy group for the user.

directoryUserId :: Maybe Text

The identifier of the user account in the directory used for identity management. If Amazon Connect cannot access the directory, you can specify this identifier to authenticate users. If you include the identifier, we assume that Amazon Connect cannot access the directory. Otherwise, the identity information is used to authenticate users from your directory.

This parameter is required if you are using an existing directory for identity management in Amazon Connect when Amazon Connect cannot access your directory to authenticate users. If you are using SAML for identity management and include this parameter, an error is returned.

password :: Maybe Text

The password for the user account. A password is required if you are using Amazon Connect for identity management. Otherwise, it is an error to include a password.

tags :: Maybe (HashMap Text Text)

One or more tags.

username :: Text

The user name for the account. For instances not using SAML for identity management, the user name can include up to 20 characters. If you are using SAML for identity management, the user name can include up to 64 characters from [a-zA-Z0-9_-.\@]+.

phoneConfig :: UserPhoneConfig

The phone settings for the user.

securityProfileIds :: NonEmpty Text

The identifier of the security profile for the user.

routingProfileId :: Text

The identifier of the routing profile for the user.

instanceId :: Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

newCreateUser

Arguments

:: Text
$sel:username:CreateUser'

-> UserPhoneConfig
$sel:phoneConfig:CreateUser'

-> NonEmpty Text
$sel:securityProfileIds:CreateUser'

-> Text
$sel:routingProfileId:CreateUser'

-> Text
$sel:instanceId:CreateUser'

-> CreateUser

Create a value of CreateUser with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identityInfo:CreateUser', createUser_identityInfo - The information about the identity of the user.

$sel:hierarchyGroupId:CreateUser', createUser_hierarchyGroupId - The identifier of the hierarchy group for the user.

$sel:directoryUserId:CreateUser', createUser_directoryUserId - The identifier of the user account in the directory used for identity management. If Amazon Connect cannot access the directory, you can specify this identifier to authenticate users. If you include the identifier, we assume that Amazon Connect cannot access the directory. Otherwise, the identity information is used to authenticate users from your directory.

This parameter is required if you are using an existing directory for identity management in Amazon Connect when Amazon Connect cannot access your directory to authenticate users. If you are using SAML for identity management and include this parameter, an error is returned.

$sel:password:CreateUser', createUser_password - The password for the user account. A password is required if you are using Amazon Connect for identity management. Otherwise, it is an error to include a password.

$sel:tags:CreateUser', createUser_tags - One or more tags.

$sel:username:CreateUser', createUser_username - The user name for the account. For instances not using SAML for identity management, the user name can include up to 20 characters. If you are using SAML for identity management, the user name can include up to 64 characters from [a-zA-Z0-9_-.\@]+.

$sel:phoneConfig:CreateUser', createUser_phoneConfig - The phone settings for the user.

$sel:securityProfileIds:CreateUser', createUser_securityProfileIds - The identifier of the security profile for the user.

$sel:routingProfileId:CreateUser', createUser_routingProfileId - The identifier of the routing profile for the user.

$sel:instanceId:CreateUser', createUser_instanceId - The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

Request Lenses

createUser_identityInfo :: Lens' CreateUser (Maybe UserIdentityInfo)

The information about the identity of the user.

createUser_hierarchyGroupId :: Lens' CreateUser (Maybe Text)

The identifier of the hierarchy group for the user.

createUser_directoryUserId :: Lens' CreateUser (Maybe Text)

The identifier of the user account in the directory used for identity management. If Amazon Connect cannot access the directory, you can specify this identifier to authenticate users. If you include the identifier, we assume that Amazon Connect cannot access the directory. Otherwise, the identity information is used to authenticate users from your directory.

This parameter is required if you are using an existing directory for identity management in Amazon Connect when Amazon Connect cannot access your directory to authenticate users. If you are using SAML for identity management and include this parameter, an error is returned.

createUser_password :: Lens' CreateUser (Maybe Text)

The password for the user account. A password is required if you are using Amazon Connect for identity management. Otherwise, it is an error to include a password.

createUser_tags :: Lens' CreateUser (Maybe (HashMap Text Text))

One or more tags.

createUser_username :: Lens' CreateUser Text

The user name for the account. For instances not using SAML for identity management, the user name can include up to 20 characters. If you are using SAML for identity management, the user name can include up to 64 characters from [a-zA-Z0-9_-.\@]+.

createUser_phoneConfig :: Lens' CreateUser UserPhoneConfig

The phone settings for the user.

createUser_securityProfileIds :: Lens' CreateUser (NonEmpty Text)

The identifier of the security profile for the user.

createUser_routingProfileId :: Lens' CreateUser Text

The identifier of the routing profile for the user.

createUser_instanceId :: Lens' CreateUser Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

Destructuring the Response

data CreateUserResponse

See: newCreateUserResponse smart constructor.

Constructors

CreateUserResponse'

Fields

userArn :: Maybe Text

The Amazon Resource Name (ARN) of the user account.

userId :: Maybe Text

The identifier of the user account.

httpStatus :: Int

The response's http status code.

newCreateUserResponse

Arguments

:: Int
$sel:httpStatus:CreateUserResponse'

-> CreateUserResponse

Create a value of CreateUserResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userArn:CreateUserResponse', createUserResponse_userArn - The Amazon Resource Name (ARN) of the user account.

$sel:userId:CreateUserResponse', createUserResponse_userId - The identifier of the user account.

$sel:httpStatus:CreateUserResponse', createUserResponse_httpStatus - The response's http status code.

Response Lenses

createUserResponse_userArn :: Lens' CreateUserResponse (Maybe Text)

The Amazon Resource Name (ARN) of the user account.

createUserResponse_userId :: Lens' CreateUserResponse (Maybe Text)

The identifier of the user account.

createUserResponse_httpStatus :: Lens' CreateUserResponse Int

The response's http status code.