Network.AWS.Connect.DescribeHoursOfOperation

Description

This API is in preview release for Amazon Connect and is subject to change.

Describes the hours of operation.

Creating a Request

data DescribeHoursOfOperation

See: newDescribeHoursOfOperation smart constructor.

Constructors

DescribeHoursOfOperation'

Fields

instanceId :: Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

hoursOfOperationId :: Text

The identifier for the hours of operation.

newDescribeHoursOfOperation

Arguments

:: Text
$sel:instanceId:DescribeHoursOfOperation'

-> Text
$sel:hoursOfOperationId:DescribeHoursOfOperation'

-> DescribeHoursOfOperation

Create a value of DescribeHoursOfOperation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceId:DescribeHoursOfOperation', describeHoursOfOperation_instanceId - The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

$sel:hoursOfOperationId:DescribeHoursOfOperation', describeHoursOfOperation_hoursOfOperationId - The identifier for the hours of operation.

Request Lenses

describeHoursOfOperation_instanceId :: Lens' DescribeHoursOfOperation Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

describeHoursOfOperation_hoursOfOperationId :: Lens' DescribeHoursOfOperation Text

The identifier for the hours of operation.

Destructuring the Response

data DescribeHoursOfOperationResponse

See: newDescribeHoursOfOperationResponse smart constructor.

Constructors

DescribeHoursOfOperationResponse'

Fields

hoursOfOperation :: Maybe HoursOfOperation

The hours of operation.

httpStatus :: Int

The response's http status code.

newDescribeHoursOfOperationResponse

Arguments

:: Int
$sel:httpStatus:DescribeHoursOfOperationResponse'

-> DescribeHoursOfOperationResponse

Create a value of DescribeHoursOfOperationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hoursOfOperation:DescribeHoursOfOperationResponse', describeHoursOfOperationResponse_hoursOfOperation - The hours of operation.

$sel:httpStatus:DescribeHoursOfOperationResponse', describeHoursOfOperationResponse_httpStatus - The response's http status code.

Response Lenses

describeHoursOfOperationResponse_hoursOfOperation :: Lens' DescribeHoursOfOperationResponse (Maybe HoursOfOperation)

The hours of operation.

describeHoursOfOperationResponse_httpStatus :: Lens' DescribeHoursOfOperationResponse Int

The response's http status code.