Network.AWS.Connect.DescribeRoutingProfile

Description

Describes the specified routing profile.

Creating a Request

data DescribeRoutingProfile

See: newDescribeRoutingProfile smart constructor.

Constructors

DescribeRoutingProfile'

Fields

instanceId :: Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

routingProfileId :: Text

The identifier of the routing profile.

newDescribeRoutingProfile

Arguments

:: Text
$sel:instanceId:DescribeRoutingProfile'

-> Text
$sel:routingProfileId:DescribeRoutingProfile'

-> DescribeRoutingProfile

Create a value of DescribeRoutingProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceId:DescribeRoutingProfile', describeRoutingProfile_instanceId - The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

$sel:routingProfileId:DescribeRoutingProfile', describeRoutingProfile_routingProfileId - The identifier of the routing profile.

Request Lenses

describeRoutingProfile_instanceId :: Lens' DescribeRoutingProfile Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

describeRoutingProfile_routingProfileId :: Lens' DescribeRoutingProfile Text

The identifier of the routing profile.

Destructuring the Response

data DescribeRoutingProfileResponse

See: newDescribeRoutingProfileResponse smart constructor.

Constructors

DescribeRoutingProfileResponse'

Fields

routingProfile :: Maybe RoutingProfile

The routing profile.

httpStatus :: Int

The response's http status code.

newDescribeRoutingProfileResponse

Arguments

:: Int
$sel:httpStatus:DescribeRoutingProfileResponse'

-> DescribeRoutingProfileResponse

Create a value of DescribeRoutingProfileResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:routingProfile:DescribeRoutingProfileResponse', describeRoutingProfileResponse_routingProfile - The routing profile.

$sel:httpStatus:DescribeRoutingProfileResponse', describeRoutingProfileResponse_httpStatus - The response's http status code.

Response Lenses

describeRoutingProfileResponse_routingProfile :: Lens' DescribeRoutingProfileResponse (Maybe RoutingProfile)

The routing profile.

describeRoutingProfileResponse_httpStatus :: Lens' DescribeRoutingProfileResponse Int

The response's http status code.