Network.AWS.Connect.DisassociateApprovedOrigin

Description

This API is in preview release for Amazon Connect and is subject to change.

Revokes access to integrated applications from Amazon Connect.

Creating a Request

data DisassociateApprovedOrigin

See: newDisassociateApprovedOrigin smart constructor.

Constructors

DisassociateApprovedOrigin'

Fields

instanceId :: Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

origin :: Text

The domain URL of the integrated application.

newDisassociateApprovedOrigin

Arguments

:: Text
$sel:instanceId:DisassociateApprovedOrigin'

-> Text
$sel:origin:DisassociateApprovedOrigin'

-> DisassociateApprovedOrigin

Create a value of DisassociateApprovedOrigin with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceId:DisassociateApprovedOrigin', disassociateApprovedOrigin_instanceId - The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

$sel:origin:DisassociateApprovedOrigin', disassociateApprovedOrigin_origin - The domain URL of the integrated application.

Request Lenses

disassociateApprovedOrigin_instanceId :: Lens' DisassociateApprovedOrigin Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

disassociateApprovedOrigin_origin :: Lens' DisassociateApprovedOrigin Text

The domain URL of the integrated application.

Destructuring the Response

data DisassociateApprovedOriginResponse

See: newDisassociateApprovedOriginResponse smart constructor.

Constructors

DisassociateApprovedOriginResponse'

newDisassociateApprovedOriginResponse :: DisassociateApprovedOriginResponse

Create a value of DisassociateApprovedOriginResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.