Network.AWS.Connect.DisassociateSecurityKey

Description

This API is in preview release for Amazon Connect and is subject to change.

Deletes the specified security key.

Creating a Request

data DisassociateSecurityKey

See: newDisassociateSecurityKey smart constructor.

Constructors

DisassociateSecurityKey'

Fields

instanceId :: Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

associationId :: Text

The existing association identifier that uniquely identifies the resource type and storage config for the given instance ID.

newDisassociateSecurityKey

Arguments

:: Text
$sel:instanceId:DisassociateSecurityKey'

-> Text
$sel:associationId:DisassociateSecurityKey'

-> DisassociateSecurityKey

Create a value of DisassociateSecurityKey with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceId:DisassociateSecurityKey', disassociateSecurityKey_instanceId - The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

$sel:associationId:DisassociateSecurityKey', disassociateSecurityKey_associationId - The existing association identifier that uniquely identifies the resource type and storage config for the given instance ID.

Request Lenses

disassociateSecurityKey_instanceId :: Lens' DisassociateSecurityKey Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

disassociateSecurityKey_associationId :: Lens' DisassociateSecurityKey Text

The existing association identifier that uniquely identifies the resource type and storage config for the given instance ID.

Destructuring the Response

data DisassociateSecurityKeyResponse

See: newDisassociateSecurityKeyResponse smart constructor.

Constructors

DisassociateSecurityKeyResponse'

newDisassociateSecurityKeyResponse :: DisassociateSecurityKeyResponse

Create a value of DisassociateSecurityKeyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.