Network.AWS.Connect.GetMetricData

Description

Gets historical metric data from the specified Amazon Connect instance.

For a description of each historical metric, see Historical Metrics Definitions in the Amazon Connect Administrator Guide.

This operation returns paginated results.

Creating a Request

data GetMetricData

See: newGetMetricData smart constructor.

Constructors

GetMetricData'

Fields

nextToken :: Maybe Text

The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

groupings :: Maybe [Grouping]

The grouping applied to the metrics returned. For example, when results are grouped by queue, the metrics returned are grouped by queue. The values returned apply to the metrics for each queue rather than aggregated for all queues.

If no grouping is specified, a summary of metrics for all queues is returned.

maxResults :: Maybe Natural

The maximum number of results to return per page.

instanceId :: Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

startTime :: POSIX

The timestamp, in UNIX Epoch time format, at which to start the reporting interval for the retrieval of historical metrics data. The time must be specified using a multiple of 5 minutes, such as 10:05, 10:10, 10:15.

The start time cannot be earlier than 24 hours before the time of the request. Historical metrics are available only for 24 hours.

endTime :: POSIX

The timestamp, in UNIX Epoch time format, at which to end the reporting interval for the retrieval of historical metrics data. The time must be specified using an interval of 5 minutes, such as 11:00, 11:05, 11:10, and must be later than the start time timestamp.

The time range between the start and end time must be less than 24 hours.

filters :: Filters

The queues, up to 100, or channels, to use to filter the metrics returned. Metric data is retrieved only for the resources associated with the queues or channels included in the filter. You can include both queue IDs and queue ARNs in the same request. VOICE, CHAT, and TASK channels are supported.

To filter by Queues, enter the queue ID/ARN, not the name of the queue.

historicalMetrics :: [HistoricalMetric]

The metrics to retrieve. Specify the name, unit, and statistic for each metric. The following historical metrics are available. For a description of each metric, see Historical Metrics Definitions in the Amazon Connect Administrator Guide.

This API does not support a contacts incoming metric (there's no CONTACTS_INCOMING metric missing from the documented list).

ABANDON_TIME

Unit: SECONDS
Statistic: AVG

AFTER_CONTACT_WORK_TIME

Unit: SECONDS
Statistic: AVG

API_CONTACTS_HANDLED

Unit: COUNT
Statistic: SUM

CALLBACK_CONTACTS_HANDLED

Unit: COUNT
Statistic: SUM

CONTACTS_ABANDONED

Unit: COUNT
Statistic: SUM

CONTACTS_AGENT_HUNG_UP_FIRST

Unit: COUNT
Statistic: SUM

CONTACTS_CONSULTED

Unit: COUNT
Statistic: SUM

CONTACTS_HANDLED

Unit: COUNT
Statistic: SUM

CONTACTS_HANDLED_INCOMING

Unit: COUNT
Statistic: SUM

CONTACTS_HANDLED_OUTBOUND

Unit: COUNT
Statistic: SUM

CONTACTS_HOLD_ABANDONS

Unit: COUNT
Statistic: SUM

CONTACTS_MISSED

Unit: COUNT
Statistic: SUM

CONTACTS_QUEUED

Unit: COUNT
Statistic: SUM

CONTACTS_TRANSFERRED_IN

Unit: COUNT
Statistic: SUM

CONTACTS_TRANSFERRED_IN_FROM_QUEUE

Unit: COUNT
Statistic: SUM

CONTACTS_TRANSFERRED_OUT

Unit: COUNT
Statistic: SUM

CONTACTS_TRANSFERRED_OUT_FROM_QUEUE

Unit: COUNT
Statistic: SUM

HANDLE_TIME

Unit: SECONDS
Statistic: AVG

HOLD_TIME

Unit: SECONDS
Statistic: AVG

INTERACTION_AND_HOLD_TIME

Unit: SECONDS
Statistic: AVG

INTERACTION_TIME

Unit: SECONDS
Statistic: AVG

OCCUPANCY

Unit: PERCENT
Statistic: AVG

QUEUE_ANSWER_TIME

Unit: SECONDS
Statistic: AVG

QUEUED_TIME

Unit: SECONDS
Statistic: MAX

SERVICE_LEVEL

You can include up to 20 SERVICE_LEVEL metrics in a request.
Unit: PERCENT

Statistic: AVG

Threshold: For ThresholdValue, enter any whole number from 1 to 604800 (inclusive), in seconds. For Comparison, you must enter LT (for "Less than").

newGetMetricData

Arguments

:: Text
$sel:instanceId:GetMetricData'

-> UTCTime
$sel:startTime:GetMetricData'

-> UTCTime
$sel:endTime:GetMetricData'

-> Filters
$sel:filters:GetMetricData'

-> GetMetricData

Create a value of GetMetricData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetMetricData', getMetricData_nextToken - The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

$sel:groupings:GetMetricData', getMetricData_groupings - The grouping applied to the metrics returned. For example, when results are grouped by queue, the metrics returned are grouped by queue. The values returned apply to the metrics for each queue rather than aggregated for all queues.

If no grouping is specified, a summary of metrics for all queues is returned.

$sel:maxResults:GetMetricData', getMetricData_maxResults - The maximum number of results to return per page.

$sel:instanceId:GetMetricData', getMetricData_instanceId - The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

$sel:startTime:GetMetricData', getMetricData_startTime - The timestamp, in UNIX Epoch time format, at which to start the reporting interval for the retrieval of historical metrics data. The time must be specified using a multiple of 5 minutes, such as 10:05, 10:10, 10:15.

The start time cannot be earlier than 24 hours before the time of the request. Historical metrics are available only for 24 hours.

$sel:endTime:GetMetricData', getMetricData_endTime - The timestamp, in UNIX Epoch time format, at which to end the reporting interval for the retrieval of historical metrics data. The time must be specified using an interval of 5 minutes, such as 11:00, 11:05, 11:10, and must be later than the start time timestamp.

The time range between the start and end time must be less than 24 hours.

$sel:filters:GetMetricData', getMetricData_filters - The queues, up to 100, or channels, to use to filter the metrics returned. Metric data is retrieved only for the resources associated with the queues or channels included in the filter. You can include both queue IDs and queue ARNs in the same request. VOICE, CHAT, and TASK channels are supported.

To filter by Queues, enter the queue ID/ARN, not the name of the queue.

$sel:historicalMetrics:GetMetricData', getMetricData_historicalMetrics - The metrics to retrieve. Specify the name, unit, and statistic for each metric. The following historical metrics are available. For a description of each metric, see Historical Metrics Definitions in the Amazon Connect Administrator Guide.

This API does not support a contacts incoming metric (there's no CONTACTS_INCOMING metric missing from the documented list).

ABANDON_TIME

Unit: SECONDS
Statistic: AVG

AFTER_CONTACT_WORK_TIME

Unit: SECONDS
Statistic: AVG

API_CONTACTS_HANDLED

Unit: COUNT
Statistic: SUM

CALLBACK_CONTACTS_HANDLED

Unit: COUNT
Statistic: SUM

CONTACTS_ABANDONED

Unit: COUNT
Statistic: SUM

CONTACTS_AGENT_HUNG_UP_FIRST

Unit: COUNT
Statistic: SUM

CONTACTS_CONSULTED

Unit: COUNT
Statistic: SUM

CONTACTS_HANDLED

Unit: COUNT
Statistic: SUM

CONTACTS_HANDLED_INCOMING

Unit: COUNT
Statistic: SUM

CONTACTS_HANDLED_OUTBOUND

Unit: COUNT
Statistic: SUM

CONTACTS_HOLD_ABANDONS

Unit: COUNT
Statistic: SUM

CONTACTS_MISSED

Unit: COUNT
Statistic: SUM

CONTACTS_QUEUED

Unit: COUNT
Statistic: SUM

CONTACTS_TRANSFERRED_IN

Unit: COUNT
Statistic: SUM

CONTACTS_TRANSFERRED_IN_FROM_QUEUE

Unit: COUNT
Statistic: SUM

CONTACTS_TRANSFERRED_OUT

Unit: COUNT
Statistic: SUM

CONTACTS_TRANSFERRED_OUT_FROM_QUEUE

Unit: COUNT
Statistic: SUM

HANDLE_TIME

Unit: SECONDS
Statistic: AVG

HOLD_TIME

Unit: SECONDS
Statistic: AVG

INTERACTION_AND_HOLD_TIME

Unit: SECONDS
Statistic: AVG

INTERACTION_TIME

Unit: SECONDS
Statistic: AVG

OCCUPANCY

Unit: PERCENT
Statistic: AVG

QUEUE_ANSWER_TIME

Unit: SECONDS
Statistic: AVG

QUEUED_TIME

Unit: SECONDS
Statistic: MAX

SERVICE_LEVEL

You can include up to 20 SERVICE_LEVEL metrics in a request.
Unit: PERCENT

Statistic: AVG

Threshold: For ThresholdValue, enter any whole number from 1 to 604800 (inclusive), in seconds. For Comparison, you must enter LT (for "Less than").

Request Lenses

getMetricData_nextToken :: Lens' GetMetricData (Maybe Text)

The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

getMetricData_groupings :: Lens' GetMetricData (Maybe [Grouping])

The grouping applied to the metrics returned. For example, when results are grouped by queue, the metrics returned are grouped by queue. The values returned apply to the metrics for each queue rather than aggregated for all queues.

If no grouping is specified, a summary of metrics for all queues is returned.

getMetricData_maxResults :: Lens' GetMetricData (Maybe Natural)

The maximum number of results to return per page.

getMetricData_instanceId :: Lens' GetMetricData Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

getMetricData_startTime :: Lens' GetMetricData UTCTime

The timestamp, in UNIX Epoch time format, at which to start the reporting interval for the retrieval of historical metrics data. The time must be specified using a multiple of 5 minutes, such as 10:05, 10:10, 10:15.

The start time cannot be earlier than 24 hours before the time of the request. Historical metrics are available only for 24 hours.

getMetricData_endTime :: Lens' GetMetricData UTCTime

The timestamp, in UNIX Epoch time format, at which to end the reporting interval for the retrieval of historical metrics data. The time must be specified using an interval of 5 minutes, such as 11:00, 11:05, 11:10, and must be later than the start time timestamp.

The time range between the start and end time must be less than 24 hours.

getMetricData_filters :: Lens' GetMetricData Filters

The queues, up to 100, or channels, to use to filter the metrics returned. Metric data is retrieved only for the resources associated with the queues or channels included in the filter. You can include both queue IDs and queue ARNs in the same request. VOICE, CHAT, and TASK channels are supported.

To filter by Queues, enter the queue ID/ARN, not the name of the queue.

getMetricData_historicalMetrics :: Lens' GetMetricData [HistoricalMetric]

The metrics to retrieve. Specify the name, unit, and statistic for each metric. The following historical metrics are available. For a description of each metric, see Historical Metrics Definitions in the Amazon Connect Administrator Guide.

This API does not support a contacts incoming metric (there's no CONTACTS_INCOMING metric missing from the documented list).

ABANDON_TIME

Unit: SECONDS
Statistic: AVG

AFTER_CONTACT_WORK_TIME

Unit: SECONDS
Statistic: AVG

API_CONTACTS_HANDLED

Unit: COUNT
Statistic: SUM

CALLBACK_CONTACTS_HANDLED

Unit: COUNT
Statistic: SUM

CONTACTS_ABANDONED

Unit: COUNT
Statistic: SUM

CONTACTS_AGENT_HUNG_UP_FIRST

Unit: COUNT
Statistic: SUM

CONTACTS_CONSULTED

Unit: COUNT
Statistic: SUM

CONTACTS_HANDLED

Unit: COUNT
Statistic: SUM

CONTACTS_HANDLED_INCOMING

Unit: COUNT
Statistic: SUM

CONTACTS_HANDLED_OUTBOUND

Unit: COUNT
Statistic: SUM

CONTACTS_HOLD_ABANDONS

Unit: COUNT
Statistic: SUM

CONTACTS_MISSED

Unit: COUNT
Statistic: SUM

CONTACTS_QUEUED

Unit: COUNT
Statistic: SUM

CONTACTS_TRANSFERRED_IN

Unit: COUNT
Statistic: SUM

CONTACTS_TRANSFERRED_IN_FROM_QUEUE

Unit: COUNT
Statistic: SUM

CONTACTS_TRANSFERRED_OUT

Unit: COUNT
Statistic: SUM

CONTACTS_TRANSFERRED_OUT_FROM_QUEUE

Unit: COUNT
Statistic: SUM

HANDLE_TIME

Unit: SECONDS
Statistic: AVG

HOLD_TIME

Unit: SECONDS
Statistic: AVG

INTERACTION_AND_HOLD_TIME

Unit: SECONDS
Statistic: AVG

INTERACTION_TIME

Unit: SECONDS
Statistic: AVG

OCCUPANCY

Unit: PERCENT
Statistic: AVG

QUEUE_ANSWER_TIME

Unit: SECONDS
Statistic: AVG

QUEUED_TIME

Unit: SECONDS
Statistic: MAX

SERVICE_LEVEL

You can include up to 20 SERVICE_LEVEL metrics in a request.
Unit: PERCENT

Statistic: AVG

Threshold: For ThresholdValue, enter any whole number from 1 to 604800 (inclusive), in seconds. For Comparison, you must enter LT (for "Less than").

Destructuring the Response

data GetMetricDataResponse

See: newGetMetricDataResponse smart constructor.

Constructors

GetMetricDataResponse'

Fields

nextToken :: Maybe Text

If there are additional results, this is the token for the next set of results.

The token expires after 5 minutes from the time it is created. Subsequent requests that use the token must use the same request parameters as the request that generated the token.

metricResults :: Maybe [HistoricalMetricResult]

Information about the historical metrics.

If no grouping is specified, a summary of metric data is returned.

httpStatus :: Int

The response's http status code.

newGetMetricDataResponse

Arguments

:: Int
$sel:httpStatus:GetMetricDataResponse'

-> GetMetricDataResponse

Create a value of GetMetricDataResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetMetricData', getMetricDataResponse_nextToken - If there are additional results, this is the token for the next set of results.

The token expires after 5 minutes from the time it is created. Subsequent requests that use the token must use the same request parameters as the request that generated the token.

$sel:metricResults:GetMetricDataResponse', getMetricDataResponse_metricResults - Information about the historical metrics.

If no grouping is specified, a summary of metric data is returned.

$sel:httpStatus:GetMetricDataResponse', getMetricDataResponse_httpStatus - The response's http status code.

Response Lenses

getMetricDataResponse_nextToken :: Lens' GetMetricDataResponse (Maybe Text)

If there are additional results, this is the token for the next set of results.

The token expires after 5 minutes from the time it is created. Subsequent requests that use the token must use the same request parameters as the request that generated the token.

getMetricDataResponse_metricResults :: Lens' GetMetricDataResponse (Maybe [HistoricalMetricResult])

Information about the historical metrics.

If no grouping is specified, a summary of metric data is returned.

getMetricDataResponse_httpStatus :: Lens' GetMetricDataResponse Int

The response's http status code.