Network.AWS.Connect.ListContactFlows

Description

Provides information about the contact flows for the specified Amazon Connect instance.

You can also create and update contact flows using the Amazon Connect Flow language.

For more information about contact flows, see Contact Flows in the Amazon Connect Administrator Guide.

This operation returns paginated results.

Creating a Request

data ListContactFlows

See: newListContactFlows smart constructor.

Constructors

ListContactFlows'

Fields

nextToken :: Maybe Text

The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

contactFlowTypes :: Maybe [ContactFlowType]

The type of contact flow.

maxResults :: Maybe Natural

The maximum number of results to return per page.

instanceId :: Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

newListContactFlows

Arguments

:: Text
$sel:instanceId:ListContactFlows'

-> ListContactFlows

Create a value of ListContactFlows with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListContactFlows', listContactFlows_nextToken - The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

$sel:contactFlowTypes:ListContactFlows', listContactFlows_contactFlowTypes - The type of contact flow.

$sel:maxResults:ListContactFlows', listContactFlows_maxResults - The maximum number of results to return per page.

$sel:instanceId:ListContactFlows', listContactFlows_instanceId - The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

Request Lenses

listContactFlows_nextToken :: Lens' ListContactFlows (Maybe Text)

The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

listContactFlows_contactFlowTypes :: Lens' ListContactFlows (Maybe [ContactFlowType])

The type of contact flow.

listContactFlows_maxResults :: Lens' ListContactFlows (Maybe Natural)

The maximum number of results to return per page.

listContactFlows_instanceId :: Lens' ListContactFlows Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

Destructuring the Response

data ListContactFlowsResponse

See: newListContactFlowsResponse smart constructor.

Constructors

ListContactFlowsResponse'

Fields

nextToken :: Maybe Text

If there are additional results, this is the token for the next set of results.

contactFlowSummaryList :: Maybe [ContactFlowSummary]

Information about the contact flows.

httpStatus :: Int

The response's http status code.

newListContactFlowsResponse

Arguments

:: Int
$sel:httpStatus:ListContactFlowsResponse'

-> ListContactFlowsResponse

Create a value of ListContactFlowsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListContactFlows', listContactFlowsResponse_nextToken - If there are additional results, this is the token for the next set of results.

$sel:contactFlowSummaryList:ListContactFlowsResponse', listContactFlowsResponse_contactFlowSummaryList - Information about the contact flows.

$sel:httpStatus:ListContactFlowsResponse', listContactFlowsResponse_httpStatus - The response's http status code.

Response Lenses

listContactFlowsResponse_nextToken :: Lens' ListContactFlowsResponse (Maybe Text)

If there are additional results, this is the token for the next set of results.

listContactFlowsResponse_contactFlowSummaryList :: Lens' ListContactFlowsResponse (Maybe [ContactFlowSummary])

Information about the contact flows.

listContactFlowsResponse_httpStatus :: Lens' ListContactFlowsResponse Int

The response's http status code.