Network.AWS.Connect.ListLexBots

Description

This API is in preview release for Amazon Connect and is subject to change.

Returns a paginated list of all the Amazon Lex bots currently associated with the instance.

This operation returns paginated results.

Creating a Request

data ListLexBots

See: newListLexBots smart constructor.

Constructors

ListLexBots'

Fields

nextToken :: Maybe Text

The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

maxResults :: Maybe Natural

The maximum number of results to return per page.

instanceId :: Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

newListLexBots

Arguments

:: Text
$sel:instanceId:ListLexBots'

-> ListLexBots

Create a value of ListLexBots with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListLexBots', listLexBots_nextToken - The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

$sel:maxResults:ListLexBots', listLexBots_maxResults - The maximum number of results to return per page.

$sel:instanceId:ListLexBots', listLexBots_instanceId - The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

Request Lenses

listLexBots_nextToken :: Lens' ListLexBots (Maybe Text)

The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

listLexBots_maxResults :: Lens' ListLexBots (Maybe Natural)

The maximum number of results to return per page.

listLexBots_instanceId :: Lens' ListLexBots Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

Destructuring the Response

data ListLexBotsResponse

See: newListLexBotsResponse smart constructor.

Constructors

ListLexBotsResponse'

Fields

nextToken :: Maybe Text

If there are additional results, this is the token for the next set of results.

lexBots :: Maybe [LexBot]

The names and Regions of the Amazon Lex bots associated with the specified instance.

httpStatus :: Int

The response's http status code.

newListLexBotsResponse

Arguments

:: Int
$sel:httpStatus:ListLexBotsResponse'

-> ListLexBotsResponse

Create a value of ListLexBotsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListLexBots', listLexBotsResponse_nextToken - If there are additional results, this is the token for the next set of results.

$sel:lexBots:ListLexBotsResponse', listLexBotsResponse_lexBots - The names and Regions of the Amazon Lex bots associated with the specified instance.

$sel:httpStatus:ListLexBotsResponse', listLexBotsResponse_httpStatus - The response's http status code.

Response Lenses

listLexBotsResponse_nextToken :: Lens' ListLexBotsResponse (Maybe Text)

If there are additional results, this is the token for the next set of results.

listLexBotsResponse_lexBots :: Lens' ListLexBotsResponse (Maybe [LexBot])

The names and Regions of the Amazon Lex bots associated with the specified instance.

listLexBotsResponse_httpStatus :: Lens' ListLexBotsResponse Int

The response's http status code.