Network.AWS.Connect.ListQueueQuickConnects

Description

This API is in preview release for Amazon Connect and is subject to change.

Lists the quick connects associated with a queue.

This operation returns paginated results.

Creating a Request

data ListQueueQuickConnects

See: newListQueueQuickConnects smart constructor.

Constructors

ListQueueQuickConnects'

Fields

nextToken :: Maybe Text

The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

maxResults :: Maybe Natural

The maximum number of results to return per page.

instanceId :: Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

queueId :: Text

The identifier for the queue.

newListQueueQuickConnects

Arguments

:: Text
$sel:instanceId:ListQueueQuickConnects'

-> Text
$sel:queueId:ListQueueQuickConnects'

-> ListQueueQuickConnects

Create a value of ListQueueQuickConnects with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListQueueQuickConnects', listQueueQuickConnects_nextToken - The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

$sel:maxResults:ListQueueQuickConnects', listQueueQuickConnects_maxResults - The maximum number of results to return per page.

$sel:instanceId:ListQueueQuickConnects', listQueueQuickConnects_instanceId - The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

$sel:queueId:ListQueueQuickConnects', listQueueQuickConnects_queueId - The identifier for the queue.

Request Lenses

listQueueQuickConnects_nextToken :: Lens' ListQueueQuickConnects (Maybe Text)

The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

listQueueQuickConnects_maxResults :: Lens' ListQueueQuickConnects (Maybe Natural)

The maximum number of results to return per page.

listQueueQuickConnects_instanceId :: Lens' ListQueueQuickConnects Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

listQueueQuickConnects_queueId :: Lens' ListQueueQuickConnects Text

The identifier for the queue.

Destructuring the Response

data ListQueueQuickConnectsResponse

See: newListQueueQuickConnectsResponse smart constructor.

Constructors

ListQueueQuickConnectsResponse'

Fields

nextToken :: Maybe Text

If there are additional results, this is the token for the next set of results.

quickConnectSummaryList :: Maybe [QuickConnectSummary]

Information about the quick connects.

httpStatus :: Int

The response's http status code.

newListQueueQuickConnectsResponse

Arguments

:: Int
$sel:httpStatus:ListQueueQuickConnectsResponse'

-> ListQueueQuickConnectsResponse

Create a value of ListQueueQuickConnectsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListQueueQuickConnects', listQueueQuickConnectsResponse_nextToken - If there are additional results, this is the token for the next set of results.

$sel:quickConnectSummaryList:ListQueueQuickConnectsResponse', listQueueQuickConnectsResponse_quickConnectSummaryList - Information about the quick connects.

$sel:httpStatus:ListQueueQuickConnectsResponse', listQueueQuickConnectsResponse_httpStatus - The response's http status code.

Response Lenses

listQueueQuickConnectsResponse_nextToken :: Lens' ListQueueQuickConnectsResponse (Maybe Text)

If there are additional results, this is the token for the next set of results.

listQueueQuickConnectsResponse_quickConnectSummaryList :: Lens' ListQueueQuickConnectsResponse (Maybe [QuickConnectSummary])

Information about the quick connects.

listQueueQuickConnectsResponse_httpStatus :: Lens' ListQueueQuickConnectsResponse Int

The response's http status code.