Network.AWS.Connect.ListRoutingProfileQueues

Description

Lists the queues associated with a routing profile.

This operation returns paginated results.

Creating a Request

data ListRoutingProfileQueues

See: newListRoutingProfileQueues smart constructor.

Constructors

ListRoutingProfileQueues'

Fields

nextToken :: Maybe Text

The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

maxResults :: Maybe Natural

The maximum number of results to return per page.

instanceId :: Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

routingProfileId :: Text

The identifier of the routing profile.

newListRoutingProfileQueues

Arguments

:: Text
$sel:instanceId:ListRoutingProfileQueues'

-> Text
$sel:routingProfileId:ListRoutingProfileQueues'

-> ListRoutingProfileQueues

Create a value of ListRoutingProfileQueues with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListRoutingProfileQueues', listRoutingProfileQueues_nextToken - The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

$sel:maxResults:ListRoutingProfileQueues', listRoutingProfileQueues_maxResults - The maximum number of results to return per page.

$sel:instanceId:ListRoutingProfileQueues', listRoutingProfileQueues_instanceId - The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

$sel:routingProfileId:ListRoutingProfileQueues', listRoutingProfileQueues_routingProfileId - The identifier of the routing profile.

Request Lenses

listRoutingProfileQueues_nextToken :: Lens' ListRoutingProfileQueues (Maybe Text)

The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

listRoutingProfileQueues_maxResults :: Lens' ListRoutingProfileQueues (Maybe Natural)

The maximum number of results to return per page.

listRoutingProfileQueues_instanceId :: Lens' ListRoutingProfileQueues Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

listRoutingProfileQueues_routingProfileId :: Lens' ListRoutingProfileQueues Text

The identifier of the routing profile.

Destructuring the Response

data ListRoutingProfileQueuesResponse

See: newListRoutingProfileQueuesResponse smart constructor.

Constructors

ListRoutingProfileQueuesResponse'

Fields

nextToken :: Maybe Text

If there are additional results, this is the token for the next set of results.

routingProfileQueueConfigSummaryList :: Maybe [RoutingProfileQueueConfigSummary]

Information about the routing profiles.

httpStatus :: Int

The response's http status code.

newListRoutingProfileQueuesResponse

Arguments

:: Int
$sel:httpStatus:ListRoutingProfileQueuesResponse'

-> ListRoutingProfileQueuesResponse

Create a value of ListRoutingProfileQueuesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListRoutingProfileQueues', listRoutingProfileQueuesResponse_nextToken - If there are additional results, this is the token for the next set of results.

$sel:routingProfileQueueConfigSummaryList:ListRoutingProfileQueuesResponse', listRoutingProfileQueuesResponse_routingProfileQueueConfigSummaryList - Information about the routing profiles.

$sel:httpStatus:ListRoutingProfileQueuesResponse', listRoutingProfileQueuesResponse_httpStatus - The response's http status code.

Response Lenses

listRoutingProfileQueuesResponse_nextToken :: Lens' ListRoutingProfileQueuesResponse (Maybe Text)

If there are additional results, this is the token for the next set of results.

listRoutingProfileQueuesResponse_routingProfileQueueConfigSummaryList :: Lens' ListRoutingProfileQueuesResponse (Maybe [RoutingProfileQueueConfigSummary])

Information about the routing profiles.

listRoutingProfileQueuesResponse_httpStatus :: Lens' ListRoutingProfileQueuesResponse Int

The response's http status code.