Network.AWS.Connect.StartContactRecording

Description

Starts recording the contact when the agent joins the call. StartContactRecording is a one-time action. For example, if you use StopContactRecording to stop recording an ongoing call, you can't use StartContactRecording to restart it. For scenarios where the recording has started and you want to suspend and resume it, such as when collecting sensitive information (for example, a credit card number), use SuspendContactRecording and ResumeContactRecording.

You can use this API to override the recording behavior configured in the Set recording behavior block.

Only voice recordings are supported at this time.

Creating a Request

data StartContactRecording

See: newStartContactRecording smart constructor.

Constructors

StartContactRecording'

Fields

instanceId :: Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

contactId :: Text

The identifier of the contact.

initialContactId :: Text

The identifier of the contact. This is the identifier of the contact associated with the first interaction with the contact center.

voiceRecordingConfiguration :: VoiceRecordingConfiguration

The person being recorded.

newStartContactRecording

Arguments

:: Text
$sel:instanceId:StartContactRecording'

-> Text
$sel:contactId:StartContactRecording'

-> Text
$sel:initialContactId:StartContactRecording'

-> VoiceRecordingConfiguration
$sel:voiceRecordingConfiguration:StartContactRecording'

-> StartContactRecording

Create a value of StartContactRecording with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceId:StartContactRecording', startContactRecording_instanceId - The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

$sel:contactId:StartContactRecording', startContactRecording_contactId - The identifier of the contact.

$sel:initialContactId:StartContactRecording', startContactRecording_initialContactId - The identifier of the contact. This is the identifier of the contact associated with the first interaction with the contact center.

$sel:voiceRecordingConfiguration:StartContactRecording', startContactRecording_voiceRecordingConfiguration - The person being recorded.

Request Lenses

startContactRecording_instanceId :: Lens' StartContactRecording Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

startContactRecording_contactId :: Lens' StartContactRecording Text

The identifier of the contact.

startContactRecording_initialContactId :: Lens' StartContactRecording Text

The identifier of the contact. This is the identifier of the contact associated with the first interaction with the contact center.

startContactRecording_voiceRecordingConfiguration :: Lens' StartContactRecording VoiceRecordingConfiguration

The person being recorded.

Destructuring the Response

data StartContactRecordingResponse

See: newStartContactRecordingResponse smart constructor.

Constructors

StartContactRecordingResponse'

Fields

httpStatus :: Int

The response's http status code.

newStartContactRecordingResponse

Arguments

:: Int
$sel:httpStatus:StartContactRecordingResponse'

-> StartContactRecordingResponse

Create a value of StartContactRecordingResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:StartContactRecordingResponse', startContactRecordingResponse_httpStatus - The response's http status code.

Response Lenses

startContactRecordingResponse_httpStatus :: Lens' StartContactRecordingResponse Int

The response's http status code.