Network.AWS.Connect.StartOutboundVoiceContact

Description

Places an outbound call to a contact, and then initiates the contact flow. It performs the actions in the contact flow that's specified (in ContactFlowId).

Agents do not initiate the outbound API, which means that they do not dial the contact. If the contact flow places an outbound call to a contact, and then puts the contact in queue, the call is then routed to the agent, like any other inbound case.

There is a 60-second dialing timeout for this operation. If the call is not connected after 60 seconds, it fails.

UK numbers with a 447 prefix are not allowed by default. Before you can dial these UK mobile numbers, you must submit a service quota increase request. For more information, see Amazon Connect Service Quotas in the Amazon Connect Administrator Guide.

Campaign calls are not allowed by default. Before you can make a call with TrafficType = CAMPAIGN, you must submit a service quota increase request. For more information, see Amazon Connect Service Quotas in the Amazon Connect Administrator Guide.

Creating a Request

data StartOutboundVoiceContact

See: newStartOutboundVoiceContact smart constructor.

Constructors

StartOutboundVoiceContact'

Fields

answerMachineDetectionConfig :: Maybe AnswerMachineDetectionConfig

Configuration of the answering machine detection for this outbound call.

queueId :: Maybe Text

The queue for the call. If you specify a queue, the phone displayed for caller ID is the phone number specified in the queue. If you do not specify a queue, the queue defined in the contact flow is used. If you do not specify a queue, you must specify a source phone number.

campaignId :: Maybe Text

The campaign identifier of the outbound communication.

trafficType :: Maybe TrafficType

Denotes the class of traffic. Calls with different traffic types are handled differently by Amazon Connect. The default value is GENERAL. Use CAMPAIGN if EnableAnswerMachineDetection is set to true. For all other cases, use GENERAL.

sourcePhoneNumber :: Maybe Text

The phone number associated with the Amazon Connect instance, in E.164 format. If you do not specify a source phone number, you must specify a queue.

attributes :: Maybe (HashMap Text Text)

A custom key-value pair using an attribute map. The attributes are standard Amazon Connect attributes, and can be accessed in contact flows just like any other contact attributes.

There can be up to 32,768 UTF-8 bytes across all key-value pairs per contact. Attribute keys can include only alphanumeric, dash, and underscore characters.

clientToken :: Maybe Text

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. The token is valid for 7 days after creation. If a contact is already started, the contact ID is returned.

destinationPhoneNumber :: Text

The phone number of the customer, in E.164 format.

contactFlowId :: Text

The identifier of the contact flow for the outbound call. To see the ContactFlowId in the Amazon Connect console user interface, on the navigation menu go to Routing, Contact Flows. Choose the contact flow. On the contact flow page, under the name of the contact flow, choose Show additional flow information. The ContactFlowId is the last part of the ARN, shown here in bold:

arn:aws:connect:us-west-2:xxxxxxxxxxxx:instance/xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx/contact-flow/846ec553-a005-41c0-8341-xxxxxxxxxxxx

instanceId :: Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

newStartOutboundVoiceContact

Arguments

:: Text
$sel:destinationPhoneNumber:StartOutboundVoiceContact'

-> Text
$sel:contactFlowId:StartOutboundVoiceContact'

-> Text
$sel:instanceId:StartOutboundVoiceContact'

-> StartOutboundVoiceContact

Create a value of StartOutboundVoiceContact with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:answerMachineDetectionConfig:StartOutboundVoiceContact', startOutboundVoiceContact_answerMachineDetectionConfig - Configuration of the answering machine detection for this outbound call.

$sel:queueId:StartOutboundVoiceContact', startOutboundVoiceContact_queueId - The queue for the call. If you specify a queue, the phone displayed for caller ID is the phone number specified in the queue. If you do not specify a queue, the queue defined in the contact flow is used. If you do not specify a queue, you must specify a source phone number.

$sel:campaignId:StartOutboundVoiceContact', startOutboundVoiceContact_campaignId - The campaign identifier of the outbound communication.

$sel:trafficType:StartOutboundVoiceContact', startOutboundVoiceContact_trafficType - Denotes the class of traffic. Calls with different traffic types are handled differently by Amazon Connect. The default value is GENERAL. Use CAMPAIGN if EnableAnswerMachineDetection is set to true. For all other cases, use GENERAL.

$sel:sourcePhoneNumber:StartOutboundVoiceContact', startOutboundVoiceContact_sourcePhoneNumber - The phone number associated with the Amazon Connect instance, in E.164 format. If you do not specify a source phone number, you must specify a queue.

$sel:attributes:StartOutboundVoiceContact', startOutboundVoiceContact_attributes - A custom key-value pair using an attribute map. The attributes are standard Amazon Connect attributes, and can be accessed in contact flows just like any other contact attributes.

There can be up to 32,768 UTF-8 bytes across all key-value pairs per contact. Attribute keys can include only alphanumeric, dash, and underscore characters.

$sel:clientToken:StartOutboundVoiceContact', startOutboundVoiceContact_clientToken - A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. The token is valid for 7 days after creation. If a contact is already started, the contact ID is returned.

$sel:destinationPhoneNumber:StartOutboundVoiceContact', startOutboundVoiceContact_destinationPhoneNumber - The phone number of the customer, in E.164 format.

$sel:contactFlowId:StartOutboundVoiceContact', startOutboundVoiceContact_contactFlowId - The identifier of the contact flow for the outbound call. To see the ContactFlowId in the Amazon Connect console user interface, on the navigation menu go to Routing, Contact Flows. Choose the contact flow. On the contact flow page, under the name of the contact flow, choose Show additional flow information. The ContactFlowId is the last part of the ARN, shown here in bold:

arn:aws:connect:us-west-2:xxxxxxxxxxxx:instance/xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx/contact-flow/846ec553-a005-41c0-8341-xxxxxxxxxxxx

$sel:instanceId:StartOutboundVoiceContact', startOutboundVoiceContact_instanceId - The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

Request Lenses

startOutboundVoiceContact_answerMachineDetectionConfig :: Lens' StartOutboundVoiceContact (Maybe AnswerMachineDetectionConfig)

Configuration of the answering machine detection for this outbound call.

startOutboundVoiceContact_queueId :: Lens' StartOutboundVoiceContact (Maybe Text)

The queue for the call. If you specify a queue, the phone displayed for caller ID is the phone number specified in the queue. If you do not specify a queue, the queue defined in the contact flow is used. If you do not specify a queue, you must specify a source phone number.

startOutboundVoiceContact_campaignId :: Lens' StartOutboundVoiceContact (Maybe Text)

The campaign identifier of the outbound communication.

startOutboundVoiceContact_trafficType :: Lens' StartOutboundVoiceContact (Maybe TrafficType)

Denotes the class of traffic. Calls with different traffic types are handled differently by Amazon Connect. The default value is GENERAL. Use CAMPAIGN if EnableAnswerMachineDetection is set to true. For all other cases, use GENERAL.

startOutboundVoiceContact_sourcePhoneNumber :: Lens' StartOutboundVoiceContact (Maybe Text)

The phone number associated with the Amazon Connect instance, in E.164 format. If you do not specify a source phone number, you must specify a queue.

startOutboundVoiceContact_attributes :: Lens' StartOutboundVoiceContact (Maybe (HashMap Text Text))

A custom key-value pair using an attribute map. The attributes are standard Amazon Connect attributes, and can be accessed in contact flows just like any other contact attributes.

There can be up to 32,768 UTF-8 bytes across all key-value pairs per contact. Attribute keys can include only alphanumeric, dash, and underscore characters.

startOutboundVoiceContact_clientToken :: Lens' StartOutboundVoiceContact (Maybe Text)

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. The token is valid for 7 days after creation. If a contact is already started, the contact ID is returned.

startOutboundVoiceContact_destinationPhoneNumber :: Lens' StartOutboundVoiceContact Text

The phone number of the customer, in E.164 format.

startOutboundVoiceContact_contactFlowId :: Lens' StartOutboundVoiceContact Text

The identifier of the contact flow for the outbound call. To see the ContactFlowId in the Amazon Connect console user interface, on the navigation menu go to Routing, Contact Flows. Choose the contact flow. On the contact flow page, under the name of the contact flow, choose Show additional flow information. The ContactFlowId is the last part of the ARN, shown here in bold:

arn:aws:connect:us-west-2:xxxxxxxxxxxx:instance/xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx/contact-flow/846ec553-a005-41c0-8341-xxxxxxxxxxxx

startOutboundVoiceContact_instanceId :: Lens' StartOutboundVoiceContact Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

Destructuring the Response

data StartOutboundVoiceContactResponse

See: newStartOutboundVoiceContactResponse smart constructor.

Constructors

StartOutboundVoiceContactResponse'

Fields

contactId :: Maybe Text

The identifier of this contact within the Amazon Connect instance.

httpStatus :: Int

The response's http status code.

newStartOutboundVoiceContactResponse

Arguments

:: Int
$sel:httpStatus:StartOutboundVoiceContactResponse'

-> StartOutboundVoiceContactResponse

Create a value of StartOutboundVoiceContactResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:contactId:StartOutboundVoiceContactResponse', startOutboundVoiceContactResponse_contactId - The identifier of this contact within the Amazon Connect instance.

$sel:httpStatus:StartOutboundVoiceContactResponse', startOutboundVoiceContactResponse_httpStatus - The response's http status code.

Response Lenses

startOutboundVoiceContactResponse_contactId :: Lens' StartOutboundVoiceContactResponse (Maybe Text)

The identifier of this contact within the Amazon Connect instance.

startOutboundVoiceContactResponse_httpStatus :: Lens' StartOutboundVoiceContactResponse Int

The response's http status code.