Network.AWS.Connect.Types.ContactFlow

Description

Documentation

data ContactFlow

Contains information about a contact flow.

See: newContactFlow smart constructor.

Constructors

ContactFlow'

Fields

arn :: Maybe Text

The Amazon Resource Name (ARN) of the contact flow.

id :: Maybe Text

The identifier of the contact flow.

name :: Maybe Text

The name of the contact flow.

content :: Maybe Text

The content of the contact flow.

tags :: Maybe (HashMap Text Text)

One or more tags.

description :: Maybe Text

The description of the contact flow.

type' :: Maybe ContactFlowType

The type of the contact flow. For descriptions of the available types, see Choose a Contact Flow Type in the Amazon Connect Administrator Guide.

newContactFlow :: ContactFlow

Create a value of ContactFlow with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:ContactFlow', contactFlow_arn - The Amazon Resource Name (ARN) of the contact flow.

$sel:id:ContactFlow', contactFlow_id - The identifier of the contact flow.

$sel:name:ContactFlow', contactFlow_name - The name of the contact flow.

$sel:content:ContactFlow', contactFlow_content - The content of the contact flow.

$sel:tags:ContactFlow', contactFlow_tags - One or more tags.

$sel:description:ContactFlow', contactFlow_description - The description of the contact flow.

$sel:type':ContactFlow', contactFlow_type - The type of the contact flow. For descriptions of the available types, see Choose a Contact Flow Type in the Amazon Connect Administrator Guide.

contactFlow_arn :: Lens' ContactFlow (Maybe Text)

The Amazon Resource Name (ARN) of the contact flow.

contactFlow_id :: Lens' ContactFlow (Maybe Text)

The identifier of the contact flow.

contactFlow_name :: Lens' ContactFlow (Maybe Text)

The name of the contact flow.

contactFlow_content :: Lens' ContactFlow (Maybe Text)

The content of the contact flow.

contactFlow_tags :: Lens' ContactFlow (Maybe (HashMap Text Text))

One or more tags.

contactFlow_description :: Lens' ContactFlow (Maybe Text)

The description of the contact flow.

contactFlow_type :: Lens' ContactFlow (Maybe ContactFlowType)

The type of the contact flow. For descriptions of the available types, see Choose a Contact Flow Type in the Amazon Connect Administrator Guide.