Network.AWS.Connect.Types.Credentials

Description

Documentation

data Credentials

Contains credentials to use for federation.

See: newCredentials smart constructor.

Constructors

Credentials'

Fields

refreshTokenExpiration :: Maybe POSIX

Renews the expiration timer for a generated token.

accessToken :: Maybe (Sensitive Text)

An access token generated for a federated user to access Amazon Connect.

accessTokenExpiration :: Maybe POSIX

A token generated with an expiration time for the session a user is logged in to Amazon Connect.

refreshToken :: Maybe (Sensitive Text)

Renews a token generated for a user to access the Amazon Connect instance.

newCredentials :: Credentials

Create a value of Credentials with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:refreshTokenExpiration:Credentials', credentials_refreshTokenExpiration - Renews the expiration timer for a generated token.

$sel:accessToken:Credentials', credentials_accessToken - An access token generated for a federated user to access Amazon Connect.

$sel:accessTokenExpiration:Credentials', credentials_accessTokenExpiration - A token generated with an expiration time for the session a user is logged in to Amazon Connect.

$sel:refreshToken:Credentials', credentials_refreshToken - Renews a token generated for a user to access the Amazon Connect instance.

credentials_refreshTokenExpiration :: Lens' Credentials (Maybe UTCTime)

Renews the expiration timer for a generated token.

credentials_accessToken :: Lens' Credentials (Maybe Text)

An access token generated for a federated user to access Amazon Connect.

credentials_accessTokenExpiration :: Lens' Credentials (Maybe UTCTime)

A token generated with an expiration time for the session a user is logged in to Amazon Connect.

credentials_refreshToken :: Lens' Credentials (Maybe Text)

Renews a token generated for a user to access the Amazon Connect instance.