Network.AWS.Connect.Types.CurrentMetric

Description

Documentation

data CurrentMetric

Contains information about a real-time metric. For a description of each metric, see Real-time Metrics Definitions in the Amazon Connect Administrator Guide.

See: newCurrentMetric smart constructor.

Constructors

CurrentMetric'

Fields

name :: Maybe CurrentMetricName

The name of the metric.

unit :: Maybe Unit

The unit for the metric.

newCurrentMetric :: CurrentMetric

Create a value of CurrentMetric with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:CurrentMetric', currentMetric_name - The name of the metric.

$sel:unit:CurrentMetric', currentMetric_unit - The unit for the metric.

currentMetric_name :: Lens' CurrentMetric (Maybe CurrentMetricName)

The name of the metric.

currentMetric_unit :: Lens' CurrentMetric (Maybe Unit)

The unit for the metric.