Network.AWS.Connect.Types.HierarchyGroup

Description

Documentation

data HierarchyGroup

Contains information about a hierarchy group.

See: newHierarchyGroup smart constructor.

Constructors

HierarchyGroup'

Fields

levelId :: Maybe Text

The identifier of the level in the hierarchy group.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the hierarchy group.

id :: Maybe Text

The identifier of the hierarchy group.

hierarchyPath :: Maybe HierarchyPath

Information about the levels in the hierarchy group.

name :: Maybe Text

The name of the hierarchy group.

newHierarchyGroup :: HierarchyGroup

Create a value of HierarchyGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:levelId:HierarchyGroup', hierarchyGroup_levelId - The identifier of the level in the hierarchy group.

$sel:arn:HierarchyGroup', hierarchyGroup_arn - The Amazon Resource Name (ARN) of the hierarchy group.

$sel:id:HierarchyGroup', hierarchyGroup_id - The identifier of the hierarchy group.

$sel:hierarchyPath:HierarchyGroup', hierarchyGroup_hierarchyPath - Information about the levels in the hierarchy group.

$sel:name:HierarchyGroup', hierarchyGroup_name - The name of the hierarchy group.

hierarchyGroup_levelId :: Lens' HierarchyGroup (Maybe Text)

The identifier of the level in the hierarchy group.

hierarchyGroup_arn :: Lens' HierarchyGroup (Maybe Text)

The Amazon Resource Name (ARN) of the hierarchy group.

hierarchyGroup_id :: Lens' HierarchyGroup (Maybe Text)

The identifier of the hierarchy group.

hierarchyGroup_hierarchyPath :: Lens' HierarchyGroup (Maybe HierarchyPath)

Information about the levels in the hierarchy group.

hierarchyGroup_name :: Lens' HierarchyGroup (Maybe Text)

The name of the hierarchy group.