Network.AWS.Connect.Types.HoursOfOperationTimeSlice

Description

Documentation

data HoursOfOperationTimeSlice

The start time or end time for an hours of operation.

See: newHoursOfOperationTimeSlice smart constructor.

Constructors

HoursOfOperationTimeSlice'

Fields

hours :: Natural

The hours.

minutes :: Natural

The minutes.

newHoursOfOperationTimeSlice

Arguments

:: Natural
$sel:hours:HoursOfOperationTimeSlice'

-> Natural
$sel:minutes:HoursOfOperationTimeSlice'

-> HoursOfOperationTimeSlice

Create a value of HoursOfOperationTimeSlice with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hours:HoursOfOperationTimeSlice', hoursOfOperationTimeSlice_hours - The hours.

$sel:minutes:HoursOfOperationTimeSlice', hoursOfOperationTimeSlice_minutes - The minutes.

hoursOfOperationTimeSlice_hours :: Lens' HoursOfOperationTimeSlice Natural

The hours.

hoursOfOperationTimeSlice_minutes :: Lens' HoursOfOperationTimeSlice Natural

The minutes.