Network.AWS.Connect.Types.LexBotConfig

Description

Documentation

data LexBotConfig

Configuration information of an Amazon Lex or Amazon Lex V2 bot.

See: newLexBotConfig smart constructor.

Constructors

LexBotConfig'

Fields

lexBot :: Maybe LexBot

lexV2Bot :: Maybe LexV2Bot

Configuration information of an Amazon Lex V2 bot.

newLexBotConfig :: LexBotConfig

Create a value of LexBotConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lexBot:LexBotConfig', lexBotConfig_lexBot - Undocumented member.

$sel:lexV2Bot:LexBotConfig', lexBotConfig_lexV2Bot - Configuration information of an Amazon Lex V2 bot.

lexBotConfig_lexBot :: Lens' LexBotConfig (Maybe LexBot)

Undocumented member.

lexBotConfig_lexV2Bot :: Lens' LexBotConfig (Maybe LexV2Bot)

Configuration information of an Amazon Lex V2 bot.