Network.AWS.Connect.Types.MediaConcurrency

Description

Documentation

data MediaConcurrency

Contains information about which channels are supported, and how many contacts an agent can have on a channel simultaneously.

See: newMediaConcurrency smart constructor.

Constructors

MediaConcurrency'

Fields

channel :: Channel

The channels that agents can handle in the Contact Control Panel (CCP).

concurrency :: Natural

The number of contacts an agent can have on a channel simultaneously.

Valid Range for VOICE: Minimum value of 1. Maximum value of 1.

Valid Range for CHAT: Minimum value of 1. Maximum value of 10.

Valid Range for TASK: Minimum value of 1. Maximum value of 10.

newMediaConcurrency

Arguments

:: Channel
$sel:channel:MediaConcurrency'

-> Natural
$sel:concurrency:MediaConcurrency'

-> MediaConcurrency

Create a value of MediaConcurrency with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channel:MediaConcurrency', mediaConcurrency_channel - The channels that agents can handle in the Contact Control Panel (CCP).

$sel:concurrency:MediaConcurrency', mediaConcurrency_concurrency - The number of contacts an agent can have on a channel simultaneously.

Valid Range for VOICE: Minimum value of 1. Maximum value of 1.

Valid Range for CHAT: Minimum value of 1. Maximum value of 10.

Valid Range for TASK: Minimum value of 1. Maximum value of 10.

mediaConcurrency_channel :: Lens' MediaConcurrency Channel

The channels that agents can handle in the Contact Control Panel (CCP).

mediaConcurrency_concurrency :: Lens' MediaConcurrency Natural

The number of contacts an agent can have on a channel simultaneously.

Valid Range for VOICE: Minimum value of 1. Maximum value of 1.

Valid Range for CHAT: Minimum value of 1. Maximum value of 10.

Valid Range for TASK: Minimum value of 1. Maximum value of 10.