Network.AWS.Connect.Types.PhoneNumberSummary

Description

Documentation

data PhoneNumberSummary

Contains summary information about a phone number for a contact center.

See: newPhoneNumberSummary smart constructor.

Constructors

PhoneNumberSummary'

Fields

phoneNumberType :: Maybe PhoneNumberType

The type of phone number.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the phone number.

phoneNumber :: Maybe Text

The phone number.

phoneNumberCountryCode :: Maybe PhoneNumberCountryCode

The ISO country code.

id :: Maybe Text

The identifier of the phone number.

newPhoneNumberSummary :: PhoneNumberSummary

Create a value of PhoneNumberSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:phoneNumberType:PhoneNumberSummary', phoneNumberSummary_phoneNumberType - The type of phone number.

$sel:arn:PhoneNumberSummary', phoneNumberSummary_arn - The Amazon Resource Name (ARN) of the phone number.

$sel:phoneNumber:PhoneNumberSummary', phoneNumberSummary_phoneNumber - The phone number.

$sel:phoneNumberCountryCode:PhoneNumberSummary', phoneNumberSummary_phoneNumberCountryCode - The ISO country code.

$sel:id:PhoneNumberSummary', phoneNumberSummary_id - The identifier of the phone number.

phoneNumberSummary_phoneNumberType :: Lens' PhoneNumberSummary (Maybe PhoneNumberType)

The type of phone number.

phoneNumberSummary_arn :: Lens' PhoneNumberSummary (Maybe Text)

The Amazon Resource Name (ARN) of the phone number.

phoneNumberSummary_phoneNumber :: Lens' PhoneNumberSummary (Maybe Text)

The phone number.

phoneNumberSummary_phoneNumberCountryCode :: Lens' PhoneNumberSummary (Maybe PhoneNumberCountryCode)

The ISO country code.

phoneNumberSummary_id :: Lens' PhoneNumberSummary (Maybe Text)

The identifier of the phone number.