Network.AWS.Connect.Types.RoutingProfileQueueConfigSummary

Description

Documentation

data RoutingProfileQueueConfigSummary

Contains summary information about a routing profile queue.

See: newRoutingProfileQueueConfigSummary smart constructor.

Constructors

RoutingProfileQueueConfigSummary'

Fields

queueId :: Text

The identifier for the queue.

queueArn :: Text

The Amazon Resource Name (ARN) of the queue.

queueName :: Text

The name of the queue.

priority :: Natural

The order in which contacts are to be handled for the queue. For more information, see Queues: priority and delay.

delay :: Natural

The delay, in seconds, that a contact should be in the queue before they are routed to an available agent. For more information, see Queues: priority and delay in the Amazon Connect Administrator Guide.

channel :: Channel

The channels this queue supports.

newRoutingProfileQueueConfigSummary

Arguments

:: Text
$sel:queueId:RoutingProfileQueueConfigSummary'

-> Text
$sel:queueArn:RoutingProfileQueueConfigSummary'

-> Text
$sel:queueName:RoutingProfileQueueConfigSummary'

-> Natural
$sel:priority:RoutingProfileQueueConfigSummary'

-> Natural
$sel:delay:RoutingProfileQueueConfigSummary'

-> Channel
$sel:channel:RoutingProfileQueueConfigSummary'

-> RoutingProfileQueueConfigSummary

Create a value of RoutingProfileQueueConfigSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:queueId:RoutingProfileQueueConfigSummary', routingProfileQueueConfigSummary_queueId - The identifier for the queue.

$sel:queueArn:RoutingProfileQueueConfigSummary', routingProfileQueueConfigSummary_queueArn - The Amazon Resource Name (ARN) of the queue.

$sel:queueName:RoutingProfileQueueConfigSummary', routingProfileQueueConfigSummary_queueName - The name of the queue.

$sel:priority:RoutingProfileQueueConfigSummary', routingProfileQueueConfigSummary_priority - The order in which contacts are to be handled for the queue. For more information, see Queues: priority and delay.

$sel:delay:RoutingProfileQueueConfigSummary', routingProfileQueueConfigSummary_delay - The delay, in seconds, that a contact should be in the queue before they are routed to an available agent. For more information, see Queues: priority and delay in the Amazon Connect Administrator Guide.

$sel:channel:RoutingProfileQueueConfigSummary', routingProfileQueueConfigSummary_channel - The channels this queue supports.

routingProfileQueueConfigSummary_queueId :: Lens' RoutingProfileQueueConfigSummary Text

The identifier for the queue.

routingProfileQueueConfigSummary_queueArn :: Lens' RoutingProfileQueueConfigSummary Text

The Amazon Resource Name (ARN) of the queue.

routingProfileQueueConfigSummary_queueName :: Lens' RoutingProfileQueueConfigSummary Text

The name of the queue.

routingProfileQueueConfigSummary_priority :: Lens' RoutingProfileQueueConfigSummary Natural

The order in which contacts are to be handled for the queue. For more information, see Queues: priority and delay.

routingProfileQueueConfigSummary_delay :: Lens' RoutingProfileQueueConfigSummary Natural

The delay, in seconds, that a contact should be in the queue before they are routed to an available agent. For more information, see Queues: priority and delay in the Amazon Connect Administrator Guide.

routingProfileQueueConfigSummary_channel :: Lens' RoutingProfileQueueConfigSummary Channel

The channels this queue supports.