Network.AWS.Connect.Types.SecurityKey

Description

Documentation

data SecurityKey

Configuration information of the security key.

See: newSecurityKey smart constructor.

Constructors

SecurityKey'

Fields

key :: Maybe Text

The key of the security key.

creationTime :: Maybe POSIX

When the security key was created.

associationId :: Maybe Text

The existing association identifier that uniquely identifies the resource type and storage config for the given instance ID.

newSecurityKey :: SecurityKey

Create a value of SecurityKey with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:SecurityKey', securityKey_key - The key of the security key.

$sel:creationTime:SecurityKey', securityKey_creationTime - When the security key was created.

$sel:associationId:SecurityKey', securityKey_associationId - The existing association identifier that uniquely identifies the resource type and storage config for the given instance ID.

securityKey_key :: Lens' SecurityKey (Maybe Text)

The key of the security key.

securityKey_creationTime :: Lens' SecurityKey (Maybe UTCTime)

When the security key was created.

securityKey_associationId :: Lens' SecurityKey (Maybe Text)

The existing association identifier that uniquely identifies the resource type and storage config for the given instance ID.