Network.AWS.Connect.Types.UserPhoneConfig

Description

Documentation

data UserPhoneConfig

Contains information about the phone configuration settings for a user.

See: newUserPhoneConfig smart constructor.

Constructors

UserPhoneConfig'

Fields

autoAccept :: Maybe Bool

The Auto accept setting.

afterContactWorkTimeLimit :: Maybe Natural

The After Call Work (ACW) timeout setting, in seconds.

deskPhoneNumber :: Maybe Text

The phone number for the user's desk phone.

phoneType :: PhoneType

The phone type.

newUserPhoneConfig

Arguments

:: PhoneType
$sel:phoneType:UserPhoneConfig'

-> UserPhoneConfig

Create a value of UserPhoneConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:autoAccept:UserPhoneConfig', userPhoneConfig_autoAccept - The Auto accept setting.

$sel:afterContactWorkTimeLimit:UserPhoneConfig', userPhoneConfig_afterContactWorkTimeLimit - The After Call Work (ACW) timeout setting, in seconds.

$sel:deskPhoneNumber:UserPhoneConfig', userPhoneConfig_deskPhoneNumber - The phone number for the user's desk phone.

$sel:phoneType:UserPhoneConfig', userPhoneConfig_phoneType - The phone type.

userPhoneConfig_autoAccept :: Lens' UserPhoneConfig (Maybe Bool)

The Auto accept setting.

userPhoneConfig_afterContactWorkTimeLimit :: Lens' UserPhoneConfig (Maybe Natural)

The After Call Work (ACW) timeout setting, in seconds.

userPhoneConfig_deskPhoneNumber :: Lens' UserPhoneConfig (Maybe Text)

The phone number for the user's desk phone.

userPhoneConfig_phoneType :: Lens' UserPhoneConfig PhoneType

The phone type.