Network.AWS.Connect.UpdateContactFlowName

Description

The name of the contact flow.

You can also create and update contact flows using the Amazon Connect Flow language.

Creating a Request

data UpdateContactFlowName

See: newUpdateContactFlowName smart constructor.

Constructors

UpdateContactFlowName'

Fields

name :: Maybe Text

The name of the contact flow.

description :: Maybe Text

The description of the contact flow.

instanceId :: Text

The identifier of the Amazon Connect instance.

contactFlowId :: Text

The identifier of the contact flow.

newUpdateContactFlowName

Arguments

:: Text
$sel:instanceId:UpdateContactFlowName'

-> Text
$sel:contactFlowId:UpdateContactFlowName'

-> UpdateContactFlowName

Create a value of UpdateContactFlowName with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:UpdateContactFlowName', updateContactFlowName_name - The name of the contact flow.

$sel:description:UpdateContactFlowName', updateContactFlowName_description - The description of the contact flow.

$sel:instanceId:UpdateContactFlowName', updateContactFlowName_instanceId - The identifier of the Amazon Connect instance.

$sel:contactFlowId:UpdateContactFlowName', updateContactFlowName_contactFlowId - The identifier of the contact flow.

Request Lenses

updateContactFlowName_name :: Lens' UpdateContactFlowName (Maybe Text)

The name of the contact flow.

updateContactFlowName_description :: Lens' UpdateContactFlowName (Maybe Text)

The description of the contact flow.

updateContactFlowName_instanceId :: Lens' UpdateContactFlowName Text

The identifier of the Amazon Connect instance.

updateContactFlowName_contactFlowId :: Lens' UpdateContactFlowName Text

The identifier of the contact flow.

Destructuring the Response

data UpdateContactFlowNameResponse

See: newUpdateContactFlowNameResponse smart constructor.

Constructors

UpdateContactFlowNameResponse'

newUpdateContactFlowNameResponse :: UpdateContactFlowNameResponse

Create a value of UpdateContactFlowNameResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.