Network.AWS.Connect.UpdateQueueOutboundCallerConfig

Description

This API is in preview release for Amazon Connect and is subject to change.

Updates the outbound caller ID name, number, and outbound whisper flow for a specified queue.

Creating a Request

data UpdateQueueOutboundCallerConfig

See: newUpdateQueueOutboundCallerConfig smart constructor.

Constructors

UpdateQueueOutboundCallerConfig'

Fields

instanceId :: Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

queueId :: Text

The identifier for the queue.

outboundCallerConfig :: OutboundCallerConfig

The outbound caller ID name, number, and outbound whisper flow.

newUpdateQueueOutboundCallerConfig

Arguments

:: Text
$sel:instanceId:UpdateQueueOutboundCallerConfig'

-> Text
$sel:queueId:UpdateQueueOutboundCallerConfig'

-> OutboundCallerConfig
$sel:outboundCallerConfig:UpdateQueueOutboundCallerConfig'

-> UpdateQueueOutboundCallerConfig

Create a value of UpdateQueueOutboundCallerConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceId:UpdateQueueOutboundCallerConfig', updateQueueOutboundCallerConfig_instanceId - The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

$sel:queueId:UpdateQueueOutboundCallerConfig', updateQueueOutboundCallerConfig_queueId - The identifier for the queue.

$sel:outboundCallerConfig:UpdateQueueOutboundCallerConfig', updateQueueOutboundCallerConfig_outboundCallerConfig - The outbound caller ID name, number, and outbound whisper flow.

Request Lenses

updateQueueOutboundCallerConfig_instanceId :: Lens' UpdateQueueOutboundCallerConfig Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

updateQueueOutboundCallerConfig_queueId :: Lens' UpdateQueueOutboundCallerConfig Text

The identifier for the queue.

updateQueueOutboundCallerConfig_outboundCallerConfig :: Lens' UpdateQueueOutboundCallerConfig OutboundCallerConfig

The outbound caller ID name, number, and outbound whisper flow.

Destructuring the Response

data UpdateQueueOutboundCallerConfigResponse

See: newUpdateQueueOutboundCallerConfigResponse smart constructor.

Constructors

UpdateQueueOutboundCallerConfigResponse'

newUpdateQueueOutboundCallerConfigResponse :: UpdateQueueOutboundCallerConfigResponse

Create a value of UpdateQueueOutboundCallerConfigResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.