Network.AWS.ConnectContactLens.Types.Transcript

Description

Documentation

data Transcript

A list of messages in the session.

See: newTranscript smart constructor.

Constructors

Transcript'

Fields

issuesDetected :: Maybe [IssueDetected]

List of positions where issues were detected on the transcript.

id :: Text

The identifier of the transcript.

participantId :: Text

The identifier of the participant.

participantRole :: Text

The role of participant. For example, is it a customer, agent, or system.

content :: Text

The content of the transcript.

beginOffsetMillis :: Natural

The beginning offset in the contact for this transcript.

endOffsetMillis :: Natural

The end offset in the contact for this transcript.

sentiment :: SentimentValue

The sentiment of the detected for this piece of transcript.

newTranscript

Arguments

:: Text
$sel:id:Transcript'

-> Text
$sel:participantId:Transcript'

-> Text
$sel:participantRole:Transcript'

-> Text
$sel:content:Transcript'

-> Natural
$sel:beginOffsetMillis:Transcript'

-> Natural
$sel:endOffsetMillis:Transcript'

-> SentimentValue
$sel:sentiment:Transcript'

-> Transcript

Create a value of Transcript with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:issuesDetected:Transcript', transcript_issuesDetected - List of positions where issues were detected on the transcript.

$sel:id:Transcript', transcript_id - The identifier of the transcript.

$sel:participantId:Transcript', transcript_participantId - The identifier of the participant.

$sel:participantRole:Transcript', transcript_participantRole - The role of participant. For example, is it a customer, agent, or system.

$sel:content:Transcript', transcript_content - The content of the transcript.

$sel:beginOffsetMillis:Transcript', transcript_beginOffsetMillis - The beginning offset in the contact for this transcript.

$sel:endOffsetMillis:Transcript', transcript_endOffsetMillis - The end offset in the contact for this transcript.

$sel:sentiment:Transcript', transcript_sentiment - The sentiment of the detected for this piece of transcript.

transcript_issuesDetected :: Lens' Transcript (Maybe [IssueDetected])

List of positions where issues were detected on the transcript.

transcript_id :: Lens' Transcript Text

The identifier of the transcript.

transcript_participantId :: Lens' Transcript Text

The identifier of the participant.

transcript_participantRole :: Lens' Transcript Text

The role of participant. For example, is it a customer, agent, or system.

transcript_content :: Lens' Transcript Text

The content of the transcript.

transcript_beginOffsetMillis :: Lens' Transcript Natural

The beginning offset in the contact for this transcript.

transcript_endOffsetMillis :: Lens' Transcript Natural

The end offset in the contact for this transcript.

transcript_sentiment :: Lens' Transcript SentimentValue

The sentiment of the detected for this piece of transcript.