Network.AWS.ConnectParticipant.SendEvent

Description

Sends an event. Note that ConnectionToken is used for invoking this API instead of ParticipantToken.

The Amazon Connect Participant Service APIs do not use Signature Version 4 authentication.

Creating a Request

data SendEvent

See: newSendEvent smart constructor.

Constructors

SendEvent'

Fields

clientToken :: Maybe Text

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

content :: Maybe Text

The content of the event to be sent (for example, message text). This is not yet supported.

contentType :: Text

The content type of the request. Supported types are:

application/vnd.amazonaws.connect.event.typing

application/vnd.amazonaws.connect.event.connection.acknowledged

connectionToken :: Text

The authentication token associated with the participant's connection.

newSendEvent

Arguments

:: Text
$sel:contentType:SendEvent'

-> Text
$sel:connectionToken:SendEvent'

-> SendEvent

Create a value of SendEvent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:SendEvent', sendEvent_clientToken - A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

$sel:content:SendEvent', sendEvent_content - The content of the event to be sent (for example, message text). This is not yet supported.

$sel:contentType:SendEvent', sendEvent_contentType - The content type of the request. Supported types are:

application/vnd.amazonaws.connect.event.typing

application/vnd.amazonaws.connect.event.connection.acknowledged

$sel:connectionToken:SendEvent', sendEvent_connectionToken - The authentication token associated with the participant's connection.

Request Lenses

sendEvent_clientToken :: Lens' SendEvent (Maybe Text)

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

sendEvent_content :: Lens' SendEvent (Maybe Text)

The content of the event to be sent (for example, message text). This is not yet supported.

sendEvent_contentType :: Lens' SendEvent Text

The content type of the request. Supported types are:

application/vnd.amazonaws.connect.event.typing

application/vnd.amazonaws.connect.event.connection.acknowledged

sendEvent_connectionToken :: Lens' SendEvent Text

The authentication token associated with the participant's connection.

Destructuring the Response

data SendEventResponse

See: newSendEventResponse smart constructor.

Constructors

SendEventResponse'

Fields

absoluteTime :: Maybe Text

The time when the event was sent.

It's specified in ISO 8601 format: yyyy-MM-ddThh:mm:ss.SSSZ. For example, 2019-11-08T02:41:28.172Z.

id :: Maybe Text

The ID of the response.

httpStatus :: Int

The response's http status code.

newSendEventResponse

Arguments

:: Int
$sel:httpStatus:SendEventResponse'

-> SendEventResponse

Create a value of SendEventResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:absoluteTime:SendEventResponse', sendEventResponse_absoluteTime - The time when the event was sent.

It's specified in ISO 8601 format: yyyy-MM-ddThh:mm:ss.SSSZ. For example, 2019-11-08T02:41:28.172Z.

$sel:id:SendEventResponse', sendEventResponse_id - The ID of the response.

$sel:httpStatus:SendEventResponse', sendEventResponse_httpStatus - The response's http status code.

Response Lenses

sendEventResponse_absoluteTime :: Lens' SendEventResponse (Maybe Text)

The time when the event was sent.

It's specified in ISO 8601 format: yyyy-MM-ddThh:mm:ss.SSSZ. For example, 2019-11-08T02:41:28.172Z.

sendEventResponse_id :: Lens' SendEventResponse (Maybe Text)

The ID of the response.

sendEventResponse_httpStatus :: Lens' SendEventResponse Int

The response's http status code.