Network.AWS.ConnectParticipant.SendMessage

Description

Sends a message. Note that ConnectionToken is used for invoking this API instead of ParticipantToken.

The Amazon Connect Participant Service APIs do not use Signature Version 4 authentication.

Creating a Request

data SendMessage

See: newSendMessage smart constructor.

Constructors

SendMessage'

Fields

clientToken :: Maybe Text

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

contentType :: Text

The type of the content. Supported types are text/plain.

content :: Text

The content of the message.

connectionToken :: Text

The authentication token associated with the connection.

newSendMessage

Arguments

:: Text
$sel:contentType:SendMessage'

-> Text
$sel:content:SendMessage'

-> Text
$sel:connectionToken:SendMessage'

-> SendMessage

Create a value of SendMessage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:SendMessage', sendMessage_clientToken - A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

$sel:contentType:SendMessage', sendMessage_contentType - The type of the content. Supported types are text/plain.

$sel:content:SendMessage', sendMessage_content - The content of the message.

$sel:connectionToken:SendMessage', sendMessage_connectionToken - The authentication token associated with the connection.

Request Lenses

sendMessage_clientToken :: Lens' SendMessage (Maybe Text)

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

sendMessage_contentType :: Lens' SendMessage Text

The type of the content. Supported types are text/plain.

sendMessage_content :: Lens' SendMessage Text

The content of the message.

sendMessage_connectionToken :: Lens' SendMessage Text

The authentication token associated with the connection.

Destructuring the Response

data SendMessageResponse

See: newSendMessageResponse smart constructor.

Constructors

SendMessageResponse'

Fields

absoluteTime :: Maybe Text

The time when the message was sent.

It's specified in ISO 8601 format: yyyy-MM-ddThh:mm:ss.SSSZ. For example, 2019-11-08T02:41:28.172Z.

id :: Maybe Text

The ID of the message.

httpStatus :: Int

The response's http status code.

newSendMessageResponse

Arguments

:: Int
$sel:httpStatus:SendMessageResponse'

-> SendMessageResponse

Create a value of SendMessageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:absoluteTime:SendMessageResponse', sendMessageResponse_absoluteTime - The time when the message was sent.

It's specified in ISO 8601 format: yyyy-MM-ddThh:mm:ss.SSSZ. For example, 2019-11-08T02:41:28.172Z.

$sel:id:SendMessageResponse', sendMessageResponse_id - The ID of the message.

$sel:httpStatus:SendMessageResponse', sendMessageResponse_httpStatus - The response's http status code.

Response Lenses

sendMessageResponse_absoluteTime :: Lens' SendMessageResponse (Maybe Text)

The time when the message was sent.

It's specified in ISO 8601 format: yyyy-MM-ddThh:mm:ss.SSSZ. For example, 2019-11-08T02:41:28.172Z.

sendMessageResponse_id :: Lens' SendMessageResponse (Maybe Text)

The ID of the message.

sendMessageResponse_httpStatus :: Lens' SendMessageResponse Int

The response's http status code.