Network.AWS.ConnectParticipant.StartAttachmentUpload

Description

Provides a pre-signed Amazon S3 URL in response for uploading the file directly to S3.

Creating a Request

data StartAttachmentUpload

See: newStartAttachmentUpload smart constructor.

Constructors

StartAttachmentUpload'

Fields

contentType :: Text

Describes the MIME file type of the attachment. For a list of supported file types, see Feature specifications in the Amazon Connect Administrator Guide.

attachmentSizeInBytes :: Natural

The size of the attachment in bytes.

attachmentName :: Text

A case-sensitive name of the attachment being uploaded.

clientToken :: Text

A unique case sensitive identifier to support idempotency of request.

connectionToken :: Text

The authentication token associated with the participant's connection.

newStartAttachmentUpload

Arguments

:: Text
$sel:contentType:StartAttachmentUpload'

-> Natural
$sel:attachmentSizeInBytes:StartAttachmentUpload'

-> Text
$sel:attachmentName:StartAttachmentUpload'

-> Text
$sel:clientToken:StartAttachmentUpload'

-> Text
$sel:connectionToken:StartAttachmentUpload'

-> StartAttachmentUpload

Create a value of StartAttachmentUpload with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:contentType:StartAttachmentUpload', startAttachmentUpload_contentType - Describes the MIME file type of the attachment. For a list of supported file types, see Feature specifications in the Amazon Connect Administrator Guide.

$sel:attachmentSizeInBytes:StartAttachmentUpload', startAttachmentUpload_attachmentSizeInBytes - The size of the attachment in bytes.

$sel:attachmentName:StartAttachmentUpload', startAttachmentUpload_attachmentName - A case-sensitive name of the attachment being uploaded.

$sel:clientToken:StartAttachmentUpload', startAttachmentUpload_clientToken - A unique case sensitive identifier to support idempotency of request.

$sel:connectionToken:StartAttachmentUpload', startAttachmentUpload_connectionToken - The authentication token associated with the participant's connection.

Request Lenses

startAttachmentUpload_contentType :: Lens' StartAttachmentUpload Text

Describes the MIME file type of the attachment. For a list of supported file types, see Feature specifications in the Amazon Connect Administrator Guide.

startAttachmentUpload_attachmentSizeInBytes :: Lens' StartAttachmentUpload Natural

The size of the attachment in bytes.

startAttachmentUpload_attachmentName :: Lens' StartAttachmentUpload Text

A case-sensitive name of the attachment being uploaded.

startAttachmentUpload_clientToken :: Lens' StartAttachmentUpload Text

A unique case sensitive identifier to support idempotency of request.

startAttachmentUpload_connectionToken :: Lens' StartAttachmentUpload Text

The authentication token associated with the participant's connection.

Destructuring the Response

data StartAttachmentUploadResponse

See: newStartAttachmentUploadResponse smart constructor.

Constructors

StartAttachmentUploadResponse'

Fields

attachmentId :: Maybe Text

A unique identifier for the attachment.

uploadMetadata :: Maybe UploadMetadata

Fields to be used while uploading the attachment.

httpStatus :: Int

The response's http status code.

newStartAttachmentUploadResponse

Arguments

:: Int
$sel:httpStatus:StartAttachmentUploadResponse'

-> StartAttachmentUploadResponse

Create a value of StartAttachmentUploadResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attachmentId:StartAttachmentUploadResponse', startAttachmentUploadResponse_attachmentId - A unique identifier for the attachment.

$sel:uploadMetadata:StartAttachmentUploadResponse', startAttachmentUploadResponse_uploadMetadata - Fields to be used while uploading the attachment.

$sel:httpStatus:StartAttachmentUploadResponse', startAttachmentUploadResponse_httpStatus - The response's http status code.

Response Lenses

startAttachmentUploadResponse_attachmentId :: Lens' StartAttachmentUploadResponse (Maybe Text)

A unique identifier for the attachment.

startAttachmentUploadResponse_uploadMetadata :: Lens' StartAttachmentUploadResponse (Maybe UploadMetadata)

Fields to be used while uploading the attachment.

startAttachmentUploadResponse_httpStatus :: Lens' StartAttachmentUploadResponse Int

The response's http status code.