Network.AWS.ConnectParticipant.Types.StartPosition

Description

Documentation

data StartPosition

A filtering option for where to start. For example, if you sent 100 messages, start with message 50.

See: newStartPosition smart constructor.

Constructors

StartPosition'

Fields

absoluteTime :: Maybe Text

The time in ISO format where to start.

It's specified in ISO 8601 format: yyyy-MM-ddThh:mm:ss.SSSZ. For example, 2019-11-08T02:41:28.172Z.

id :: Maybe Text

The ID of the message or event where to start.

mostRecent :: Maybe Natural

The start position of the most recent message where you want to start.

newStartPosition :: StartPosition

Create a value of StartPosition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:absoluteTime:StartPosition', startPosition_absoluteTime - The time in ISO format where to start.

It's specified in ISO 8601 format: yyyy-MM-ddThh:mm:ss.SSSZ. For example, 2019-11-08T02:41:28.172Z.

$sel:id:StartPosition', startPosition_id - The ID of the message or event where to start.

$sel:mostRecent:StartPosition', startPosition_mostRecent - The start position of the most recent message where you want to start.

startPosition_absoluteTime :: Lens' StartPosition (Maybe Text)

The time in ISO format where to start.

It's specified in ISO 8601 format: yyyy-MM-ddThh:mm:ss.SSSZ. For example, 2019-11-08T02:41:28.172Z.

startPosition_id :: Lens' StartPosition (Maybe Text)

The ID of the message or event where to start.

startPosition_mostRecent :: Lens' StartPosition (Maybe Natural)

The start position of the most recent message where you want to start.