Network.AWS.CostAndUsageReport.Types

Description

Service Configuration

defaultService :: Service

API version 2017-01-06 of the Amazon Cost and Usage Report Service SDK configuration.

Errors

_DuplicateReportNameException :: AsError a => Getting (First ServiceError) a ServiceError

A report with the specified name already exists in the account. Specify a different report name.

_InternalErrorException :: AsError a => Getting (First ServiceError) a ServiceError

An error on the server occurred during the processing of your request. Try again later.

_ReportLimitReachedException :: AsError a => Getting (First ServiceError) a ServiceError

This account already has five reports defined. To define a new report, you must delete an existing report.

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The input fails to satisfy the constraints specified by an AWS service.

AWSRegion

newtype AWSRegion

The region of the S3 bucket that AWS delivers the report into.

Constructors

AWSRegion'

Fields

fromAWSRegion :: Text

Bundled Patterns

pattern AWSRegion_Af_south_1 :: AWSRegion

pattern AWSRegion_Ap_east_1 :: AWSRegion

pattern AWSRegion_Ap_northeast_1 :: AWSRegion

pattern AWSRegion_Ap_northeast_2 :: AWSRegion

pattern AWSRegion_Ap_northeast_3 :: AWSRegion

pattern AWSRegion_Ap_south_1 :: AWSRegion

pattern AWSRegion_Ap_southeast_1 :: AWSRegion

pattern AWSRegion_Ap_southeast_2 :: AWSRegion

pattern AWSRegion_Ca_central_1 :: AWSRegion

pattern AWSRegion_Cn_north_1 :: AWSRegion

pattern AWSRegion_Cn_northwest_1 :: AWSRegion

pattern AWSRegion_Eu_central_1 :: AWSRegion

pattern AWSRegion_Eu_north_1 :: AWSRegion

pattern AWSRegion_Eu_south_1 :: AWSRegion

pattern AWSRegion_Eu_west_1 :: AWSRegion

pattern AWSRegion_Eu_west_2 :: AWSRegion

pattern AWSRegion_Eu_west_3 :: AWSRegion

pattern AWSRegion_Me_south_1 :: AWSRegion

pattern AWSRegion_Sa_east_1 :: AWSRegion

pattern AWSRegion_Us_east_1 :: AWSRegion

pattern AWSRegion_Us_east_2 :: AWSRegion

pattern AWSRegion_Us_west_1 :: AWSRegion

pattern AWSRegion_Us_west_2 :: AWSRegion

AdditionalArtifact

newtype AdditionalArtifact

The types of manifest that you want AWS to create for this report.

Constructors

AdditionalArtifact'

Fields

fromAdditionalArtifact :: Text

Bundled Patterns

pattern AdditionalArtifact_ATHENA :: AdditionalArtifact

pattern AdditionalArtifact_QUICKSIGHT :: AdditionalArtifact

pattern AdditionalArtifact_REDSHIFT :: AdditionalArtifact

CompressionFormat

newtype CompressionFormat

The compression format that AWS uses for the report.

Constructors

CompressionFormat'

Fields

fromCompressionFormat :: Text

Bundled Patterns

pattern CompressionFormat_GZIP :: CompressionFormat

pattern CompressionFormat_Parquet :: CompressionFormat

pattern CompressionFormat_ZIP :: CompressionFormat

ReportFormat

newtype ReportFormat

The format that AWS saves the report in.

Constructors

ReportFormat'

Fields

fromReportFormat :: Text

Bundled Patterns

pattern ReportFormat_Parquet :: ReportFormat

pattern ReportFormat_TextORcsv :: ReportFormat

ReportVersioning

newtype ReportVersioning

Constructors

ReportVersioning'

Fields

fromReportVersioning :: Text

Bundled Patterns

pattern ReportVersioning_CREATE_NEW_REPORT :: ReportVersioning

pattern ReportVersioning_OVERWRITE_REPORT :: ReportVersioning

SchemaElement

newtype SchemaElement

Whether or not AWS includes resource IDs in the report.

Constructors

SchemaElement'

Fields

fromSchemaElement :: Text

Bundled Patterns

pattern SchemaElement_RESOURCES :: SchemaElement

TimeUnit

newtype TimeUnit

The length of time covered by the report.

Constructors

TimeUnit'

Fields

fromTimeUnit :: Text

Bundled Patterns

pattern TimeUnit_DAILY :: TimeUnit

pattern TimeUnit_HOURLY :: TimeUnit

pattern TimeUnit_MONTHLY :: TimeUnit

ReportDefinition

data ReportDefinition

The definition of AWS Cost and Usage Report. You can specify the report name, time unit, report format, compression format, S3 bucket, additional artifacts, and schema elements in the definition.

See: newReportDefinition smart constructor.

Constructors

ReportDefinition'

Fields

reportVersioning :: Maybe ReportVersioning

Whether you want Amazon Web Services to overwrite the previous version of each report or to deliver the report in addition to the previous versions.

additionalArtifacts :: Maybe [AdditionalArtifact]

A list of manifests that you want Amazon Web Services to create for this report.

billingViewArn :: Maybe Text

The Amazon resource name of the billing view. You can get this value by using the billing view service public APIs.

refreshClosedReports :: Maybe Bool

Whether you want Amazon Web Services to update your reports after they have been finalized if Amazon Web Services detects charges related to previous months. These charges can include refunds, credits, or support fees.

reportName :: Text

timeUnit :: TimeUnit

format :: ReportFormat

compression :: CompressionFormat

additionalSchemaElements :: [SchemaElement]

A list of strings that indicate additional content that Amazon Web Services includes in the report, such as individual resource IDs.

s3Bucket :: Text

s3Prefix :: Text

s3Region :: AWSRegion

newReportDefinition

Arguments

:: Text
$sel:reportName:ReportDefinition'

-> TimeUnit
$sel:timeUnit:ReportDefinition'

-> ReportFormat
$sel:format:ReportDefinition'

-> CompressionFormat
$sel:compression:ReportDefinition'

-> Text
$sel:s3Bucket:ReportDefinition'

-> Text
$sel:s3Prefix:ReportDefinition'

-> AWSRegion
$sel:s3Region:ReportDefinition'

-> ReportDefinition

Create a value of ReportDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reportVersioning:ReportDefinition', reportDefinition_reportVersioning - Whether you want Amazon Web Services to overwrite the previous version of each report or to deliver the report in addition to the previous versions.

$sel:additionalArtifacts:ReportDefinition', reportDefinition_additionalArtifacts - A list of manifests that you want Amazon Web Services to create for this report.

$sel:billingViewArn:ReportDefinition', reportDefinition_billingViewArn - The Amazon resource name of the billing view. You can get this value by using the billing view service public APIs.

$sel:refreshClosedReports:ReportDefinition', reportDefinition_refreshClosedReports - Whether you want Amazon Web Services to update your reports after they have been finalized if Amazon Web Services detects charges related to previous months. These charges can include refunds, credits, or support fees.

$sel:reportName:ReportDefinition', reportDefinition_reportName - Undocumented member.

$sel:timeUnit:ReportDefinition', reportDefinition_timeUnit - Undocumented member.

$sel:format:ReportDefinition', reportDefinition_format - Undocumented member.

$sel:compression:ReportDefinition', reportDefinition_compression - Undocumented member.

$sel:additionalSchemaElements:ReportDefinition', reportDefinition_additionalSchemaElements - A list of strings that indicate additional content that Amazon Web Services includes in the report, such as individual resource IDs.

$sel:s3Bucket:ReportDefinition', reportDefinition_s3Bucket - Undocumented member.

$sel:s3Prefix:ReportDefinition', reportDefinition_s3Prefix - Undocumented member.

$sel:s3Region:ReportDefinition', reportDefinition_s3Region - Undocumented member.

reportDefinition_reportVersioning :: Lens' ReportDefinition (Maybe ReportVersioning)

Whether you want Amazon Web Services to overwrite the previous version of each report or to deliver the report in addition to the previous versions.

reportDefinition_additionalArtifacts :: Lens' ReportDefinition (Maybe [AdditionalArtifact])

A list of manifests that you want Amazon Web Services to create for this report.

reportDefinition_billingViewArn :: Lens' ReportDefinition (Maybe Text)

The Amazon resource name of the billing view. You can get this value by using the billing view service public APIs.

reportDefinition_refreshClosedReports :: Lens' ReportDefinition (Maybe Bool)

Whether you want Amazon Web Services to update your reports after they have been finalized if Amazon Web Services detects charges related to previous months. These charges can include refunds, credits, or support fees.

reportDefinition_reportName :: Lens' ReportDefinition Text

Undocumented member.

reportDefinition_timeUnit :: Lens' ReportDefinition TimeUnit

Undocumented member.

reportDefinition_format :: Lens' ReportDefinition ReportFormat

Undocumented member.

reportDefinition_compression :: Lens' ReportDefinition CompressionFormat

Undocumented member.

reportDefinition_additionalSchemaElements :: Lens' ReportDefinition [SchemaElement]

A list of strings that indicate additional content that Amazon Web Services includes in the report, such as individual resource IDs.

reportDefinition_s3Bucket :: Lens' ReportDefinition Text

Undocumented member.

reportDefinition_s3Prefix :: Lens' ReportDefinition Text

Undocumented member.

reportDefinition_s3Region :: Lens' ReportDefinition AWSRegion

Undocumented member.