Network.AWS.CostAndUsageReport.Types.ReportDefinition

Description

Documentation

data ReportDefinition

The definition of AWS Cost and Usage Report. You can specify the report name, time unit, report format, compression format, S3 bucket, additional artifacts, and schema elements in the definition.

See: newReportDefinition smart constructor.

Constructors

ReportDefinition'

Fields

reportVersioning :: Maybe ReportVersioning

Whether you want Amazon Web Services to overwrite the previous version of each report or to deliver the report in addition to the previous versions.

additionalArtifacts :: Maybe [AdditionalArtifact]

A list of manifests that you want Amazon Web Services to create for this report.

billingViewArn :: Maybe Text

The Amazon resource name of the billing view. You can get this value by using the billing view service public APIs.

refreshClosedReports :: Maybe Bool

Whether you want Amazon Web Services to update your reports after they have been finalized if Amazon Web Services detects charges related to previous months. These charges can include refunds, credits, or support fees.

reportName :: Text

timeUnit :: TimeUnit

format :: ReportFormat

compression :: CompressionFormat

additionalSchemaElements :: [SchemaElement]

A list of strings that indicate additional content that Amazon Web Services includes in the report, such as individual resource IDs.

s3Bucket :: Text

s3Prefix :: Text

s3Region :: AWSRegion

newReportDefinition

Arguments

:: Text
$sel:reportName:ReportDefinition'

-> TimeUnit
$sel:timeUnit:ReportDefinition'

-> ReportFormat
$sel:format:ReportDefinition'

-> CompressionFormat
$sel:compression:ReportDefinition'

-> Text
$sel:s3Bucket:ReportDefinition'

-> Text
$sel:s3Prefix:ReportDefinition'

-> AWSRegion
$sel:s3Region:ReportDefinition'

-> ReportDefinition

Create a value of ReportDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reportVersioning:ReportDefinition', reportDefinition_reportVersioning - Whether you want Amazon Web Services to overwrite the previous version of each report or to deliver the report in addition to the previous versions.

$sel:additionalArtifacts:ReportDefinition', reportDefinition_additionalArtifacts - A list of manifests that you want Amazon Web Services to create for this report.

$sel:billingViewArn:ReportDefinition', reportDefinition_billingViewArn - The Amazon resource name of the billing view. You can get this value by using the billing view service public APIs.

$sel:refreshClosedReports:ReportDefinition', reportDefinition_refreshClosedReports - Whether you want Amazon Web Services to update your reports after they have been finalized if Amazon Web Services detects charges related to previous months. These charges can include refunds, credits, or support fees.

$sel:reportName:ReportDefinition', reportDefinition_reportName - Undocumented member.

$sel:timeUnit:ReportDefinition', reportDefinition_timeUnit - Undocumented member.

$sel:format:ReportDefinition', reportDefinition_format - Undocumented member.

$sel:compression:ReportDefinition', reportDefinition_compression - Undocumented member.

$sel:additionalSchemaElements:ReportDefinition', reportDefinition_additionalSchemaElements - A list of strings that indicate additional content that Amazon Web Services includes in the report, such as individual resource IDs.

$sel:s3Bucket:ReportDefinition', reportDefinition_s3Bucket - Undocumented member.

$sel:s3Prefix:ReportDefinition', reportDefinition_s3Prefix - Undocumented member.

$sel:s3Region:ReportDefinition', reportDefinition_s3Region - Undocumented member.

reportDefinition_reportVersioning :: Lens' ReportDefinition (Maybe ReportVersioning)

Whether you want Amazon Web Services to overwrite the previous version of each report or to deliver the report in addition to the previous versions.

reportDefinition_additionalArtifacts :: Lens' ReportDefinition (Maybe [AdditionalArtifact])

A list of manifests that you want Amazon Web Services to create for this report.

reportDefinition_billingViewArn :: Lens' ReportDefinition (Maybe Text)

The Amazon resource name of the billing view. You can get this value by using the billing view service public APIs.

reportDefinition_refreshClosedReports :: Lens' ReportDefinition (Maybe Bool)

Whether you want Amazon Web Services to update your reports after they have been finalized if Amazon Web Services detects charges related to previous months. These charges can include refunds, credits, or support fees.

reportDefinition_reportName :: Lens' ReportDefinition Text

Undocumented member.

reportDefinition_timeUnit :: Lens' ReportDefinition TimeUnit

Undocumented member.

reportDefinition_format :: Lens' ReportDefinition ReportFormat

Undocumented member.

reportDefinition_compression :: Lens' ReportDefinition CompressionFormat

Undocumented member.

reportDefinition_additionalSchemaElements :: Lens' ReportDefinition [SchemaElement]

A list of strings that indicate additional content that Amazon Web Services includes in the report, such as individual resource IDs.

reportDefinition_s3Bucket :: Lens' ReportDefinition Text

Undocumented member.

reportDefinition_s3Prefix :: Lens' ReportDefinition Text

Undocumented member.

reportDefinition_s3Region :: Lens' ReportDefinition AWSRegion

Undocumented member.