Network.AWS.CostExplorer.CreateCostCategoryDefinition

Description

Creates a new Cost Category with the requested name and rules.

Creating a Request

data CreateCostCategoryDefinition

See: newCreateCostCategoryDefinition smart constructor.

Constructors

CreateCostCategoryDefinition'

Fields

splitChargeRules :: Maybe (NonEmpty CostCategorySplitChargeRule)

The split charge rules used to allocate your charges between your Cost Category values.

defaultValue :: Maybe Text

name :: Text

ruleVersion :: CostCategoryRuleVersion

rules :: NonEmpty CostCategoryRule

The Cost Category rules used to categorize costs. For more information, see CostCategoryRule.

newCreateCostCategoryDefinition

Arguments

:: Text
$sel:name:CreateCostCategoryDefinition'

-> CostCategoryRuleVersion
$sel:ruleVersion:CreateCostCategoryDefinition'

-> NonEmpty CostCategoryRule
$sel:rules:CreateCostCategoryDefinition'

-> CreateCostCategoryDefinition

Create a value of CreateCostCategoryDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:splitChargeRules:CreateCostCategoryDefinition', createCostCategoryDefinition_splitChargeRules - The split charge rules used to allocate your charges between your Cost Category values.

$sel:defaultValue:CreateCostCategoryDefinition', createCostCategoryDefinition_defaultValue - Undocumented member.

$sel:name:CreateCostCategoryDefinition', createCostCategoryDefinition_name - Undocumented member.

$sel:ruleVersion:CreateCostCategoryDefinition', createCostCategoryDefinition_ruleVersion - Undocumented member.

$sel:rules:CreateCostCategoryDefinition', createCostCategoryDefinition_rules - The Cost Category rules used to categorize costs. For more information, see CostCategoryRule.

Request Lenses

createCostCategoryDefinition_splitChargeRules :: Lens' CreateCostCategoryDefinition (Maybe (NonEmpty CostCategorySplitChargeRule))

The split charge rules used to allocate your charges between your Cost Category values.

createCostCategoryDefinition_defaultValue :: Lens' CreateCostCategoryDefinition (Maybe Text)

Undocumented member.

createCostCategoryDefinition_name :: Lens' CreateCostCategoryDefinition Text

Undocumented member.

createCostCategoryDefinition_ruleVersion :: Lens' CreateCostCategoryDefinition CostCategoryRuleVersion

Undocumented member.

createCostCategoryDefinition_rules :: Lens' CreateCostCategoryDefinition (NonEmpty CostCategoryRule)

The Cost Category rules used to categorize costs. For more information, see CostCategoryRule.

Destructuring the Response

data CreateCostCategoryDefinitionResponse

See: newCreateCostCategoryDefinitionResponse smart constructor.

Constructors

CreateCostCategoryDefinitionResponse'

Fields

costCategoryArn :: Maybe Text

The unique identifier for your newly created Cost Category.

effectiveStart :: Maybe Text

The Cost Category's effective start date.

httpStatus :: Int

The response's http status code.

newCreateCostCategoryDefinitionResponse

Arguments

:: Int
$sel:httpStatus:CreateCostCategoryDefinitionResponse'

-> CreateCostCategoryDefinitionResponse

Create a value of CreateCostCategoryDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:costCategoryArn:CreateCostCategoryDefinitionResponse', createCostCategoryDefinitionResponse_costCategoryArn - The unique identifier for your newly created Cost Category.

$sel:effectiveStart:CreateCostCategoryDefinitionResponse', createCostCategoryDefinitionResponse_effectiveStart - The Cost Category's effective start date.

$sel:httpStatus:CreateCostCategoryDefinitionResponse', createCostCategoryDefinitionResponse_httpStatus - The response's http status code.

Response Lenses

createCostCategoryDefinitionResponse_costCategoryArn :: Lens' CreateCostCategoryDefinitionResponse (Maybe Text)

The unique identifier for your newly created Cost Category.

createCostCategoryDefinitionResponse_effectiveStart :: Lens' CreateCostCategoryDefinitionResponse (Maybe Text)

The Cost Category's effective start date.

createCostCategoryDefinitionResponse_httpStatus :: Lens' CreateCostCategoryDefinitionResponse Int

The response's http status code.