Network.AWS.CostExplorer.DeleteAnomalySubscription

Description

Deletes a cost anomaly subscription.

Creating a Request

data DeleteAnomalySubscription

See: newDeleteAnomalySubscription smart constructor.

Constructors

DeleteAnomalySubscription'

Fields

subscriptionArn :: Text

The unique identifier of the cost anomaly subscription that you want to delete.

newDeleteAnomalySubscription

Arguments

:: Text
$sel:subscriptionArn:DeleteAnomalySubscription'

-> DeleteAnomalySubscription

Create a value of DeleteAnomalySubscription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:subscriptionArn:DeleteAnomalySubscription', deleteAnomalySubscription_subscriptionArn - The unique identifier of the cost anomaly subscription that you want to delete.

Request Lenses

deleteAnomalySubscription_subscriptionArn :: Lens' DeleteAnomalySubscription Text

The unique identifier of the cost anomaly subscription that you want to delete.

Destructuring the Response

data DeleteAnomalySubscriptionResponse

See: newDeleteAnomalySubscriptionResponse smart constructor.

Constructors

DeleteAnomalySubscriptionResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteAnomalySubscriptionResponse

Arguments

:: Int
$sel:httpStatus:DeleteAnomalySubscriptionResponse'

-> DeleteAnomalySubscriptionResponse

Create a value of DeleteAnomalySubscriptionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteAnomalySubscriptionResponse', deleteAnomalySubscriptionResponse_httpStatus - The response's http status code.

Response Lenses

deleteAnomalySubscriptionResponse_httpStatus :: Lens' DeleteAnomalySubscriptionResponse Int

The response's http status code.