Network.AWS.CostExplorer.DeleteCostCategoryDefinition

Description

Deletes a Cost Category. Expenses from this month going forward will no longer be categorized with this Cost Category.

Creating a Request

data DeleteCostCategoryDefinition

See: newDeleteCostCategoryDefinition smart constructor.

Constructors

DeleteCostCategoryDefinition'

Fields

costCategoryArn :: Text

The unique identifier for your Cost Category.

newDeleteCostCategoryDefinition

Arguments

:: Text
$sel:costCategoryArn:DeleteCostCategoryDefinition'

-> DeleteCostCategoryDefinition

Create a value of DeleteCostCategoryDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:costCategoryArn:DeleteCostCategoryDefinition', deleteCostCategoryDefinition_costCategoryArn - The unique identifier for your Cost Category.

Request Lenses

deleteCostCategoryDefinition_costCategoryArn :: Lens' DeleteCostCategoryDefinition Text

The unique identifier for your Cost Category.

Destructuring the Response

data DeleteCostCategoryDefinitionResponse

See: newDeleteCostCategoryDefinitionResponse smart constructor.

Constructors

DeleteCostCategoryDefinitionResponse'

Fields

costCategoryArn :: Maybe Text

The unique identifier for your Cost Category.

effectiveEnd :: Maybe Text

The effective end date of the Cost Category as a result of deleting it. No costs after this date will be categorized by the deleted Cost Category.

httpStatus :: Int

The response's http status code.

newDeleteCostCategoryDefinitionResponse

Arguments

:: Int
$sel:httpStatus:DeleteCostCategoryDefinitionResponse'

-> DeleteCostCategoryDefinitionResponse

Create a value of DeleteCostCategoryDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:costCategoryArn:DeleteCostCategoryDefinition', deleteCostCategoryDefinitionResponse_costCategoryArn - The unique identifier for your Cost Category.

$sel:effectiveEnd:DeleteCostCategoryDefinitionResponse', deleteCostCategoryDefinitionResponse_effectiveEnd - The effective end date of the Cost Category as a result of deleting it. No costs after this date will be categorized by the deleted Cost Category.

$sel:httpStatus:DeleteCostCategoryDefinitionResponse', deleteCostCategoryDefinitionResponse_httpStatus - The response's http status code.

Response Lenses

deleteCostCategoryDefinitionResponse_costCategoryArn :: Lens' DeleteCostCategoryDefinitionResponse (Maybe Text)

The unique identifier for your Cost Category.

deleteCostCategoryDefinitionResponse_effectiveEnd :: Lens' DeleteCostCategoryDefinitionResponse (Maybe Text)

The effective end date of the Cost Category as a result of deleting it. No costs after this date will be categorized by the deleted Cost Category.

deleteCostCategoryDefinitionResponse_httpStatus :: Lens' DeleteCostCategoryDefinitionResponse Int

The response's http status code.