Network.AWS.CostExplorer.GetCostCategories

Description

Retrieves an array of Cost Category names and values incurred cost.

If some Cost Category names and values are not associated with any cost, they will not be returned by this API.

Creating a Request

data GetCostCategories

See: newGetCostCategories smart constructor.

Constructors

GetCostCategories'

Fields

nextPageToken :: Maybe Text

If the number of objects that are still available for retrieval exceeds the limit, Amazon Web Services returns a NextPageToken value in the response. To retrieve the next batch of objects, provide the NextPageToken from the prior call in your next request.

searchString :: Maybe Text

The value that you want to search the filter values for.

If you do not specify a CostCategoryName, SearchString will be used to filter Cost Category names that match the SearchString pattern. If you do specifiy a CostCategoryName, SearchString will be used to filter Cost Category values that match the SearchString pattern.

costCategoryName :: Maybe Text

filter' :: Maybe Expression

maxResults :: Maybe Natural

This field is only used when SortBy is provided in the request.

The maximum number of objects that to be returned for this request. If MaxResults is not specified with SortBy, the request will return 1000 results as the default value for this parameter.

For GetCostCategories, MaxResults has an upper limit of 1000.

sortBy :: Maybe [SortDefinition]

The value by which you want to sort the data.

The key represents cost and usage metrics. The following values are supported:

BlendedCost

UnblendedCost

AmortizedCost

NetAmortizedCost

NetUnblendedCost

UsageQuantity

NormalizedUsageAmount

Supported values for SortOrder are ASCENDING or DESCENDING.

When using SortBy, NextPageToken and SearchString are not supported.

timePeriod :: DateInterval

newGetCostCategories

Arguments

:: DateInterval
$sel:timePeriod:GetCostCategories'

-> GetCostCategories

Create a value of GetCostCategories with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextPageToken:GetCostCategories', getCostCategories_nextPageToken - If the number of objects that are still available for retrieval exceeds the limit, Amazon Web Services returns a NextPageToken value in the response. To retrieve the next batch of objects, provide the NextPageToken from the prior call in your next request.

$sel:searchString:GetCostCategories', getCostCategories_searchString - The value that you want to search the filter values for.

If you do not specify a CostCategoryName, SearchString will be used to filter Cost Category names that match the SearchString pattern. If you do specifiy a CostCategoryName, SearchString will be used to filter Cost Category values that match the SearchString pattern.

$sel:costCategoryName:GetCostCategories', getCostCategories_costCategoryName - Undocumented member.

$sel:filter':GetCostCategories', getCostCategories_filter - Undocumented member.

$sel:maxResults:GetCostCategories', getCostCategories_maxResults - This field is only used when SortBy is provided in the request.

The maximum number of objects that to be returned for this request. If MaxResults is not specified with SortBy, the request will return 1000 results as the default value for this parameter.

For GetCostCategories, MaxResults has an upper limit of 1000.

$sel:sortBy:GetCostCategories', getCostCategories_sortBy - The value by which you want to sort the data.

The key represents cost and usage metrics. The following values are supported:

BlendedCost

UnblendedCost

AmortizedCost

NetAmortizedCost

NetUnblendedCost

UsageQuantity

NormalizedUsageAmount

Supported values for SortOrder are ASCENDING or DESCENDING.

When using SortBy, NextPageToken and SearchString are not supported.

$sel:timePeriod:GetCostCategories', getCostCategories_timePeriod - Undocumented member.

Request Lenses

getCostCategories_nextPageToken :: Lens' GetCostCategories (Maybe Text)

If the number of objects that are still available for retrieval exceeds the limit, Amazon Web Services returns a NextPageToken value in the response. To retrieve the next batch of objects, provide the NextPageToken from the prior call in your next request.

getCostCategories_searchString :: Lens' GetCostCategories (Maybe Text)

The value that you want to search the filter values for.

If you do not specify a CostCategoryName, SearchString will be used to filter Cost Category names that match the SearchString pattern. If you do specifiy a CostCategoryName, SearchString will be used to filter Cost Category values that match the SearchString pattern.

getCostCategories_costCategoryName :: Lens' GetCostCategories (Maybe Text)

Undocumented member.

getCostCategories_filter :: Lens' GetCostCategories (Maybe Expression)

Undocumented member.

getCostCategories_maxResults :: Lens' GetCostCategories (Maybe Natural)

This field is only used when SortBy is provided in the request.

The maximum number of objects that to be returned for this request. If MaxResults is not specified with SortBy, the request will return 1000 results as the default value for this parameter.

For GetCostCategories, MaxResults has an upper limit of 1000.

getCostCategories_sortBy :: Lens' GetCostCategories (Maybe [SortDefinition])

The value by which you want to sort the data.

The key represents cost and usage metrics. The following values are supported:

BlendedCost

UnblendedCost

AmortizedCost

NetAmortizedCost

NetUnblendedCost

UsageQuantity

NormalizedUsageAmount

Supported values for SortOrder are ASCENDING or DESCENDING.

When using SortBy, NextPageToken and SearchString are not supported.

getCostCategories_timePeriod :: Lens' GetCostCategories DateInterval

Undocumented member.

Destructuring the Response

data GetCostCategoriesResponse

See: newGetCostCategoriesResponse smart constructor.

Constructors

GetCostCategoriesResponse'

Fields

nextPageToken :: Maybe Text

If the number of objects that are still available for retrieval exceeds the limit, Amazon Web Services returns a NextPageToken value in the response. To retrieve the next batch of objects, provide the marker from the prior call in your next request.

costCategoryNames :: Maybe [Text]

The names of the Cost Categories.

costCategoryValues :: Maybe [Text]

The Cost Category values.

CostCategoryValues are not returned if CostCategoryName is not specified in the request.

httpStatus :: Int

The response's http status code.

returnSize :: Int

The number of objects returned.

totalSize :: Int

The total number of objects.

newGetCostCategoriesResponse

Arguments

:: Int
$sel:httpStatus:GetCostCategoriesResponse'

-> Int
$sel:returnSize:GetCostCategoriesResponse'

-> Int
$sel:totalSize:GetCostCategoriesResponse'

-> GetCostCategoriesResponse

Create a value of GetCostCategoriesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextPageToken:GetCostCategories', getCostCategoriesResponse_nextPageToken - If the number of objects that are still available for retrieval exceeds the limit, Amazon Web Services returns a NextPageToken value in the response. To retrieve the next batch of objects, provide the marker from the prior call in your next request.

$sel:costCategoryNames:GetCostCategoriesResponse', getCostCategoriesResponse_costCategoryNames - The names of the Cost Categories.

$sel:costCategoryValues:GetCostCategoriesResponse', getCostCategoriesResponse_costCategoryValues - The Cost Category values.

CostCategoryValues are not returned if CostCategoryName is not specified in the request.

$sel:httpStatus:GetCostCategoriesResponse', getCostCategoriesResponse_httpStatus - The response's http status code.

$sel:returnSize:GetCostCategoriesResponse', getCostCategoriesResponse_returnSize - The number of objects returned.

$sel:totalSize:GetCostCategoriesResponse', getCostCategoriesResponse_totalSize - The total number of objects.

Response Lenses

getCostCategoriesResponse_nextPageToken :: Lens' GetCostCategoriesResponse (Maybe Text)

If the number of objects that are still available for retrieval exceeds the limit, Amazon Web Services returns a NextPageToken value in the response. To retrieve the next batch of objects, provide the marker from the prior call in your next request.

getCostCategoriesResponse_costCategoryNames :: Lens' GetCostCategoriesResponse (Maybe [Text])

The names of the Cost Categories.

getCostCategoriesResponse_costCategoryValues :: Lens' GetCostCategoriesResponse (Maybe [Text])

The Cost Category values.

CostCategoryValues are not returned if CostCategoryName is not specified in the request.

getCostCategoriesResponse_httpStatus :: Lens' GetCostCategoriesResponse Int

The response's http status code.

getCostCategoriesResponse_returnSize :: Lens' GetCostCategoriesResponse Int

The number of objects returned.

getCostCategoriesResponse_totalSize :: Lens' GetCostCategoriesResponse Int

The total number of objects.