Network.AWS.CostExplorer.GetCostForecast

Description

Retrieves a forecast for how much Amazon Web Services predicts that you will spend over the forecast time period that you select, based on your past costs.

Creating a Request

data GetCostForecast

See: newGetCostForecast smart constructor.

Constructors

GetCostForecast'

Fields

predictionIntervalLevel :: Maybe Natural

Cost Explorer always returns the mean forecast as a single point. You can request a prediction interval around the mean by specifying a confidence level. The higher the confidence level, the more confident Cost Explorer is about the actual value falling in the prediction interval. Higher confidence levels result in wider prediction intervals.

filter' :: Maybe Expression

The filters that you want to use to filter your forecast. The GetCostForecast API supports filtering by the following dimensions:

AZ

INSTANCE_TYPE

LINKED_ACCOUNT

LINKED_ACCOUNT_NAME

OPERATION

PURCHASE_TYPE

REGION

SERVICE

USAGE_TYPE

USAGE_TYPE_GROUP

RECORD_TYPE

OPERATING_SYSTEM

TENANCY

SCOPE

PLATFORM

SUBSCRIPTION_ID

LEGAL_ENTITY_NAME

DEPLOYMENT_OPTION

DATABASE_ENGINE

INSTANCE_TYPE_FAMILY

BILLING_ENTITY

RESERVATION_ID

SAVINGS_PLAN_ARN

timePeriod :: DateInterval

The period of time that you want the forecast to cover. The start date must be equal to or no later than the current date to avoid a validation error.

metric :: Metric

Which metric Cost Explorer uses to create your forecast. For more information about blended and unblended rates, see Why does the "blended" annotation appear on some line items in my bill?.

Valid values for a GetCostForecast call are the following:

AMORTIZED_COST

BLENDED_COST

NET_AMORTIZED_COST

NET_UNBLENDED_COST

UNBLENDED_COST

granularity :: Granularity

How granular you want the forecast to be. You can get 3 months of DAILY forecasts or 12 months of MONTHLY forecasts.

The GetCostForecast operation supports only DAILY and MONTHLY granularities.

newGetCostForecast

Arguments

:: DateInterval
$sel:timePeriod:GetCostForecast'

-> Metric
$sel:metric:GetCostForecast'

-> Granularity
$sel:granularity:GetCostForecast'

-> GetCostForecast

Create a value of GetCostForecast with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:predictionIntervalLevel:GetCostForecast', getCostForecast_predictionIntervalLevel - Cost Explorer always returns the mean forecast as a single point. You can request a prediction interval around the mean by specifying a confidence level. The higher the confidence level, the more confident Cost Explorer is about the actual value falling in the prediction interval. Higher confidence levels result in wider prediction intervals.

$sel:filter':GetCostForecast', getCostForecast_filter - The filters that you want to use to filter your forecast. The GetCostForecast API supports filtering by the following dimensions:

AZ

INSTANCE_TYPE

LINKED_ACCOUNT

LINKED_ACCOUNT_NAME

OPERATION

PURCHASE_TYPE

REGION

SERVICE

USAGE_TYPE

USAGE_TYPE_GROUP

RECORD_TYPE

OPERATING_SYSTEM

TENANCY

SCOPE

PLATFORM

SUBSCRIPTION_ID

LEGAL_ENTITY_NAME

DEPLOYMENT_OPTION

DATABASE_ENGINE

INSTANCE_TYPE_FAMILY

BILLING_ENTITY

RESERVATION_ID

SAVINGS_PLAN_ARN

$sel:timePeriod:GetCostForecast', getCostForecast_timePeriod - The period of time that you want the forecast to cover. The start date must be equal to or no later than the current date to avoid a validation error.

$sel:metric:GetCostForecast', getCostForecast_metric - Which metric Cost Explorer uses to create your forecast. For more information about blended and unblended rates, see Why does the "blended" annotation appear on some line items in my bill?.

Valid values for a GetCostForecast call are the following:

AMORTIZED_COST

BLENDED_COST

NET_AMORTIZED_COST

NET_UNBLENDED_COST

UNBLENDED_COST

$sel:granularity:GetCostForecast', getCostForecast_granularity - How granular you want the forecast to be. You can get 3 months of DAILY forecasts or 12 months of MONTHLY forecasts.

The GetCostForecast operation supports only DAILY and MONTHLY granularities.

Request Lenses

getCostForecast_predictionIntervalLevel :: Lens' GetCostForecast (Maybe Natural)

Cost Explorer always returns the mean forecast as a single point. You can request a prediction interval around the mean by specifying a confidence level. The higher the confidence level, the more confident Cost Explorer is about the actual value falling in the prediction interval. Higher confidence levels result in wider prediction intervals.

getCostForecast_filter :: Lens' GetCostForecast (Maybe Expression)

The filters that you want to use to filter your forecast. The GetCostForecast API supports filtering by the following dimensions:

AZ

INSTANCE_TYPE

LINKED_ACCOUNT

LINKED_ACCOUNT_NAME

OPERATION

PURCHASE_TYPE

REGION

SERVICE

USAGE_TYPE

USAGE_TYPE_GROUP

RECORD_TYPE

OPERATING_SYSTEM

TENANCY

SCOPE

PLATFORM

SUBSCRIPTION_ID

LEGAL_ENTITY_NAME

DEPLOYMENT_OPTION

DATABASE_ENGINE

INSTANCE_TYPE_FAMILY

BILLING_ENTITY

RESERVATION_ID

SAVINGS_PLAN_ARN

getCostForecast_timePeriod :: Lens' GetCostForecast DateInterval

The period of time that you want the forecast to cover. The start date must be equal to or no later than the current date to avoid a validation error.

getCostForecast_metric :: Lens' GetCostForecast Metric

Which metric Cost Explorer uses to create your forecast. For more information about blended and unblended rates, see Why does the "blended" annotation appear on some line items in my bill?.

Valid values for a GetCostForecast call are the following:

AMORTIZED_COST

BLENDED_COST

NET_AMORTIZED_COST

NET_UNBLENDED_COST

UNBLENDED_COST

getCostForecast_granularity :: Lens' GetCostForecast Granularity

How granular you want the forecast to be. You can get 3 months of DAILY forecasts or 12 months of MONTHLY forecasts.

The GetCostForecast operation supports only DAILY and MONTHLY granularities.

Destructuring the Response

data GetCostForecastResponse

See: newGetCostForecastResponse smart constructor.

Constructors

GetCostForecastResponse'

Fields

forecastResultsByTime :: Maybe [ForecastResult]

The forecasts for your query, in order. For DAILY forecasts, this is a list of days. For MONTHLY forecasts, this is a list of months.

total :: Maybe MetricValue

How much you are forecasted to spend over the forecast period, in USD.

httpStatus :: Int

The response's http status code.

newGetCostForecastResponse

Arguments

:: Int
$sel:httpStatus:GetCostForecastResponse'

-> GetCostForecastResponse

Create a value of GetCostForecastResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:forecastResultsByTime:GetCostForecastResponse', getCostForecastResponse_forecastResultsByTime - The forecasts for your query, in order. For DAILY forecasts, this is a list of days. For MONTHLY forecasts, this is a list of months.

$sel:total:GetCostForecastResponse', getCostForecastResponse_total - How much you are forecasted to spend over the forecast period, in USD.

$sel:httpStatus:GetCostForecastResponse', getCostForecastResponse_httpStatus - The response's http status code.

Response Lenses

getCostForecastResponse_forecastResultsByTime :: Lens' GetCostForecastResponse (Maybe [ForecastResult])

The forecasts for your query, in order. For DAILY forecasts, this is a list of days. For MONTHLY forecasts, this is a list of months.

getCostForecastResponse_total :: Lens' GetCostForecastResponse (Maybe MetricValue)

How much you are forecasted to spend over the forecast period, in USD.

getCostForecastResponse_httpStatus :: Lens' GetCostForecastResponse Int

The response's http status code.