Network.AWS.CostExplorer.GetDimensionValues

Description

Retrieves all available filter values for a specified filter over a period of time. You can search the dimension values for an arbitrary string.

Creating a Request

data GetDimensionValues

See: newGetDimensionValues smart constructor.

Constructors

GetDimensionValues'

Fields

nextPageToken :: Maybe Text

The token to retrieve the next set of results. Amazon Web Services provides the token when the response from a previous call has more results than the maximum page size.

context :: Maybe Context

The context for the call to GetDimensionValues. This can be RESERVATIONS or COST_AND_USAGE. The default value is COST_AND_USAGE. If the context is set to RESERVATIONS, the resulting dimension values can be used in the GetReservationUtilization operation. If the context is set to COST_AND_USAGE, the resulting dimension values can be used in the GetCostAndUsage operation.

If you set the context to COST_AND_USAGE, you can use the following dimensions for searching:

AZ - The Availability Zone. An example is us-east-1a.

DATABASE_ENGINE - The Amazon Relational Database Service database. Examples are Aurora or MySQL.

INSTANCE_TYPE - The type of Amazon EC2 instance. An example is m4.xlarge.

LEGAL_ENTITY_NAME - The name of the organization that sells you Amazon Web Services services, such as Amazon Web Services.

LINKED_ACCOUNT - The description in the attribute map that includes the full name of the member account. The value field contains the Amazon Web Services ID of the member account.

OPERATING_SYSTEM - The operating system. Examples are Windows or Linux.

OPERATION - The action performed. Examples include RunInstance and CreateBucket.

PLATFORM - The Amazon EC2 operating system. Examples are Windows or Linux.

PURCHASE_TYPE - The reservation type of the purchase to which this usage is related. Examples include On-Demand Instances and Standard Reserved Instances.

SERVICE - The Amazon Web Services service such as Amazon DynamoDB.

USAGE_TYPE - The type of usage. An example is DataTransfer-In-Bytes. The response for the GetDimensionValues operation includes a unit attribute. Examples include GB and Hrs.

USAGE_TYPE_GROUP - The grouping of common usage types. An example is Amazon EC2: CloudWatch – Alarms. The response for this operation includes a unit attribute.

REGION - The Amazon Web Services Region.

RECORD_TYPE - The different types of charges such as RI fees, usage costs, tax refunds, and credits.

RESOURCE_ID - The unique identifier of the resource. ResourceId is an opt-in feature only available for last 14 days for EC2-Compute Service.

If you set the context to RESERVATIONS, you can use the following dimensions for searching:

AZ - The Availability Zone. An example is us-east-1a.

CACHE_ENGINE - The Amazon ElastiCache operating system. Examples are Windows or Linux.

DEPLOYMENT_OPTION - The scope of Amazon Relational Database Service deployments. Valid values are SingleAZ and MultiAZ.

INSTANCE_TYPE - The type of Amazon EC2 instance. An example is m4.xlarge.

LINKED_ACCOUNT - The description in the attribute map that includes the full name of the member account. The value field contains the Amazon Web Services ID of the member account.

PLATFORM - The Amazon EC2 operating system. Examples are Windows or Linux.

REGION - The Amazon Web Services Region.

SCOPE (Utilization only) - The scope of a Reserved Instance (RI). Values are regional or a single Availability Zone.

TAG (Coverage only) - The tags that are associated with a Reserved Instance (RI).

TENANCY - The tenancy of a resource. Examples are shared or dedicated.

If you set the context to SAVINGS_PLANS, you can use the following dimensions for searching:

SAVINGS_PLANS_TYPE - Type of Savings Plans (EC2 Instance or Compute)

PAYMENT_OPTION - Payment option for the given Savings Plans (for example, All Upfront)

REGION - The Amazon Web Services Region.

INSTANCE_TYPE_FAMILY - The family of instances (For example, m5)

LINKED_ACCOUNT - The description in the attribute map that includes the full name of the member account. The value field contains the Amazon Web Services ID of the member account.

SAVINGS_PLAN_ARN - The unique identifier for your Savings Plan

searchString :: Maybe Text

The value that you want to search the filter values for.

filter' :: Maybe Expression

maxResults :: Maybe Natural

This field is only used when SortBy is provided in the request. The maximum number of objects that to be returned for this request. If MaxResults is not specified with SortBy, the request will return 1000 results as the default value for this parameter.

For GetDimensionValues, MaxResults has an upper limit of 1000.

sortBy :: Maybe [SortDefinition]

The value by which you want to sort the data.

The key represents cost and usage metrics. The following values are supported:

BlendedCost

UnblendedCost

AmortizedCost

NetAmortizedCost

NetUnblendedCost

UsageQuantity

NormalizedUsageAmount

Supported values for SortOrder are ASCENDING or DESCENDING.

When you specify a SortBy paramater, the context must be COST_AND_USAGE. Further, when using SortBy, NextPageToken and SearchString are not supported.

timePeriod :: DateInterval

The start date and end date for retrieving the dimension values. The start date is inclusive, but the end date is exclusive. For example, if start is 2017-01-01 and end is 2017-05-01, then the cost and usage data is retrieved from 2017-01-01 up to and including 2017-04-30 but not including 2017-05-01.

dimension :: Dimension

The name of the dimension. Each Dimension is available for a different Context. For more information, see Context.

newGetDimensionValues

Arguments

:: DateInterval
$sel:timePeriod:GetDimensionValues'

-> Dimension
$sel:dimension:GetDimensionValues'

-> GetDimensionValues

Create a value of GetDimensionValues with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextPageToken:GetDimensionValues', getDimensionValues_nextPageToken - The token to retrieve the next set of results. Amazon Web Services provides the token when the response from a previous call has more results than the maximum page size.

$sel:context:GetDimensionValues', getDimensionValues_context - The context for the call to GetDimensionValues. This can be RESERVATIONS or COST_AND_USAGE. The default value is COST_AND_USAGE. If the context is set to RESERVATIONS, the resulting dimension values can be used in the GetReservationUtilization operation. If the context is set to COST_AND_USAGE, the resulting dimension values can be used in the GetCostAndUsage operation.

If you set the context to COST_AND_USAGE, you can use the following dimensions for searching:

AZ - The Availability Zone. An example is us-east-1a.

DATABASE_ENGINE - The Amazon Relational Database Service database. Examples are Aurora or MySQL.

INSTANCE_TYPE - The type of Amazon EC2 instance. An example is m4.xlarge.

LEGAL_ENTITY_NAME - The name of the organization that sells you Amazon Web Services services, such as Amazon Web Services.

LINKED_ACCOUNT - The description in the attribute map that includes the full name of the member account. The value field contains the Amazon Web Services ID of the member account.

OPERATING_SYSTEM - The operating system. Examples are Windows or Linux.

OPERATION - The action performed. Examples include RunInstance and CreateBucket.

PLATFORM - The Amazon EC2 operating system. Examples are Windows or Linux.

PURCHASE_TYPE - The reservation type of the purchase to which this usage is related. Examples include On-Demand Instances and Standard Reserved Instances.

SERVICE - The Amazon Web Services service such as Amazon DynamoDB.

USAGE_TYPE - The type of usage. An example is DataTransfer-In-Bytes. The response for the GetDimensionValues operation includes a unit attribute. Examples include GB and Hrs.

USAGE_TYPE_GROUP - The grouping of common usage types. An example is Amazon EC2: CloudWatch – Alarms. The response for this operation includes a unit attribute.

REGION - The Amazon Web Services Region.

RECORD_TYPE - The different types of charges such as RI fees, usage costs, tax refunds, and credits.

RESOURCE_ID - The unique identifier of the resource. ResourceId is an opt-in feature only available for last 14 days for EC2-Compute Service.

If you set the context to RESERVATIONS, you can use the following dimensions for searching:

AZ - The Availability Zone. An example is us-east-1a.

CACHE_ENGINE - The Amazon ElastiCache operating system. Examples are Windows or Linux.

DEPLOYMENT_OPTION - The scope of Amazon Relational Database Service deployments. Valid values are SingleAZ and MultiAZ.

INSTANCE_TYPE - The type of Amazon EC2 instance. An example is m4.xlarge.

LINKED_ACCOUNT - The description in the attribute map that includes the full name of the member account. The value field contains the Amazon Web Services ID of the member account.

PLATFORM - The Amazon EC2 operating system. Examples are Windows or Linux.

REGION - The Amazon Web Services Region.

SCOPE (Utilization only) - The scope of a Reserved Instance (RI). Values are regional or a single Availability Zone.

TAG (Coverage only) - The tags that are associated with a Reserved Instance (RI).

TENANCY - The tenancy of a resource. Examples are shared or dedicated.

If you set the context to SAVINGS_PLANS, you can use the following dimensions for searching:

SAVINGS_PLANS_TYPE - Type of Savings Plans (EC2 Instance or Compute)

PAYMENT_OPTION - Payment option for the given Savings Plans (for example, All Upfront)

REGION - The Amazon Web Services Region.

INSTANCE_TYPE_FAMILY - The family of instances (For example, m5)

LINKED_ACCOUNT - The description in the attribute map that includes the full name of the member account. The value field contains the Amazon Web Services ID of the member account.

SAVINGS_PLAN_ARN - The unique identifier for your Savings Plan

$sel:searchString:GetDimensionValues', getDimensionValues_searchString - The value that you want to search the filter values for.

$sel:filter':GetDimensionValues', getDimensionValues_filter - Undocumented member.

$sel:maxResults:GetDimensionValues', getDimensionValues_maxResults - This field is only used when SortBy is provided in the request. The maximum number of objects that to be returned for this request. If MaxResults is not specified with SortBy, the request will return 1000 results as the default value for this parameter.

For GetDimensionValues, MaxResults has an upper limit of 1000.

$sel:sortBy:GetDimensionValues', getDimensionValues_sortBy - The value by which you want to sort the data.

The key represents cost and usage metrics. The following values are supported:

BlendedCost

UnblendedCost

AmortizedCost

NetAmortizedCost

NetUnblendedCost

UsageQuantity

NormalizedUsageAmount

Supported values for SortOrder are ASCENDING or DESCENDING.

When you specify a SortBy paramater, the context must be COST_AND_USAGE. Further, when using SortBy, NextPageToken and SearchString are not supported.

$sel:timePeriod:GetDimensionValues', getDimensionValues_timePeriod - The start date and end date for retrieving the dimension values. The start date is inclusive, but the end date is exclusive. For example, if start is 2017-01-01 and end is 2017-05-01, then the cost and usage data is retrieved from 2017-01-01 up to and including 2017-04-30 but not including 2017-05-01.

$sel:dimension:GetDimensionValues', getDimensionValues_dimension - The name of the dimension. Each Dimension is available for a different Context. For more information, see Context.

Request Lenses

getDimensionValues_nextPageToken :: Lens' GetDimensionValues (Maybe Text)

The token to retrieve the next set of results. Amazon Web Services provides the token when the response from a previous call has more results than the maximum page size.

getDimensionValues_context :: Lens' GetDimensionValues (Maybe Context)

The context for the call to GetDimensionValues. This can be RESERVATIONS or COST_AND_USAGE. The default value is COST_AND_USAGE. If the context is set to RESERVATIONS, the resulting dimension values can be used in the GetReservationUtilization operation. If the context is set to COST_AND_USAGE, the resulting dimension values can be used in the GetCostAndUsage operation.

If you set the context to COST_AND_USAGE, you can use the following dimensions for searching:

AZ - The Availability Zone. An example is us-east-1a.

DATABASE_ENGINE - The Amazon Relational Database Service database. Examples are Aurora or MySQL.

INSTANCE_TYPE - The type of Amazon EC2 instance. An example is m4.xlarge.

LEGAL_ENTITY_NAME - The name of the organization that sells you Amazon Web Services services, such as Amazon Web Services.

LINKED_ACCOUNT - The description in the attribute map that includes the full name of the member account. The value field contains the Amazon Web Services ID of the member account.

OPERATING_SYSTEM - The operating system. Examples are Windows or Linux.

OPERATION - The action performed. Examples include RunInstance and CreateBucket.

PLATFORM - The Amazon EC2 operating system. Examples are Windows or Linux.

PURCHASE_TYPE - The reservation type of the purchase to which this usage is related. Examples include On-Demand Instances and Standard Reserved Instances.

SERVICE - The Amazon Web Services service such as Amazon DynamoDB.

USAGE_TYPE - The type of usage. An example is DataTransfer-In-Bytes. The response for the GetDimensionValues operation includes a unit attribute. Examples include GB and Hrs.

USAGE_TYPE_GROUP - The grouping of common usage types. An example is Amazon EC2: CloudWatch – Alarms. The response for this operation includes a unit attribute.

REGION - The Amazon Web Services Region.

RECORD_TYPE - The different types of charges such as RI fees, usage costs, tax refunds, and credits.

RESOURCE_ID - The unique identifier of the resource. ResourceId is an opt-in feature only available for last 14 days for EC2-Compute Service.

If you set the context to RESERVATIONS, you can use the following dimensions for searching:

AZ - The Availability Zone. An example is us-east-1a.

CACHE_ENGINE - The Amazon ElastiCache operating system. Examples are Windows or Linux.

DEPLOYMENT_OPTION - The scope of Amazon Relational Database Service deployments. Valid values are SingleAZ and MultiAZ.

INSTANCE_TYPE - The type of Amazon EC2 instance. An example is m4.xlarge.

LINKED_ACCOUNT - The description in the attribute map that includes the full name of the member account. The value field contains the Amazon Web Services ID of the member account.

PLATFORM - The Amazon EC2 operating system. Examples are Windows or Linux.

REGION - The Amazon Web Services Region.

SCOPE (Utilization only) - The scope of a Reserved Instance (RI). Values are regional or a single Availability Zone.

TAG (Coverage only) - The tags that are associated with a Reserved Instance (RI).

TENANCY - The tenancy of a resource. Examples are shared or dedicated.

If you set the context to SAVINGS_PLANS, you can use the following dimensions for searching:

SAVINGS_PLANS_TYPE - Type of Savings Plans (EC2 Instance or Compute)

PAYMENT_OPTION - Payment option for the given Savings Plans (for example, All Upfront)

REGION - The Amazon Web Services Region.

INSTANCE_TYPE_FAMILY - The family of instances (For example, m5)

LINKED_ACCOUNT - The description in the attribute map that includes the full name of the member account. The value field contains the Amazon Web Services ID of the member account.

SAVINGS_PLAN_ARN - The unique identifier for your Savings Plan

getDimensionValues_searchString :: Lens' GetDimensionValues (Maybe Text)

The value that you want to search the filter values for.

getDimensionValues_filter :: Lens' GetDimensionValues (Maybe Expression)

Undocumented member.

getDimensionValues_maxResults :: Lens' GetDimensionValues (Maybe Natural)

This field is only used when SortBy is provided in the request. The maximum number of objects that to be returned for this request. If MaxResults is not specified with SortBy, the request will return 1000 results as the default value for this parameter.

For GetDimensionValues, MaxResults has an upper limit of 1000.

getDimensionValues_sortBy :: Lens' GetDimensionValues (Maybe [SortDefinition])

The value by which you want to sort the data.

The key represents cost and usage metrics. The following values are supported:

BlendedCost

UnblendedCost

AmortizedCost

NetAmortizedCost

NetUnblendedCost

UsageQuantity

NormalizedUsageAmount

Supported values for SortOrder are ASCENDING or DESCENDING.

When you specify a SortBy paramater, the context must be COST_AND_USAGE. Further, when using SortBy, NextPageToken and SearchString are not supported.

getDimensionValues_timePeriod :: Lens' GetDimensionValues DateInterval

The start date and end date for retrieving the dimension values. The start date is inclusive, but the end date is exclusive. For example, if start is 2017-01-01 and end is 2017-05-01, then the cost and usage data is retrieved from 2017-01-01 up to and including 2017-04-30 but not including 2017-05-01.

getDimensionValues_dimension :: Lens' GetDimensionValues Dimension

The name of the dimension. Each Dimension is available for a different Context. For more information, see Context.

Destructuring the Response

data GetDimensionValuesResponse

See: newGetDimensionValuesResponse smart constructor.

Constructors

GetDimensionValuesResponse'

Fields

nextPageToken :: Maybe Text

The token for the next set of retrievable results. Amazon Web Services provides the token when the response from a previous call has more results than the maximum page size.

httpStatus :: Int

The response's http status code.

dimensionValues :: [DimensionValuesWithAttributes]

The filters that you used to filter your request. Some dimensions are available only for a specific context.

If you set the context to COST_AND_USAGE, you can use the following dimensions for searching:

AZ - The Availability Zone. An example is us-east-1a.

DATABASE_ENGINE - The Amazon Relational Database Service database. Examples are Aurora or MySQL.

INSTANCE_TYPE - The type of Amazon EC2 instance. An example is m4.xlarge.

LEGAL_ENTITY_NAME - The name of the organization that sells you Amazon Web Services services, such as Amazon Web Services.

LINKED_ACCOUNT - The description in the attribute map that includes the full name of the member account. The value field contains the Amazon Web Services ID of the member account.

OPERATING_SYSTEM - The operating system. Examples are Windows or Linux.

OPERATION - The action performed. Examples include RunInstance and CreateBucket.

PLATFORM - The Amazon EC2 operating system. Examples are Windows or Linux.

PURCHASE_TYPE - The reservation type of the purchase to which this usage is related. Examples include On-Demand Instances and Standard Reserved Instances.

SERVICE - The Amazon Web Services service such as Amazon DynamoDB.

USAGE_TYPE - The type of usage. An example is DataTransfer-In-Bytes. The response for the GetDimensionValues operation includes a unit attribute. Examples include GB and Hrs.

USAGE_TYPE_GROUP - The grouping of common usage types. An example is Amazon EC2: CloudWatch – Alarms. The response for this operation includes a unit attribute.

RECORD_TYPE - The different types of charges such as RI fees, usage costs, tax refunds, and credits.

RESOURCE_ID - The unique identifier of the resource. ResourceId is an opt-in feature only available for last 14 days for EC2-Compute Service.

If you set the context to RESERVATIONS, you can use the following dimensions for searching:

AZ - The Availability Zone. An example is us-east-1a.

CACHE_ENGINE - The Amazon ElastiCache operating system. Examples are Windows or Linux.

DEPLOYMENT_OPTION - The scope of Amazon Relational Database Service deployments. Valid values are SingleAZ and MultiAZ.

INSTANCE_TYPE - The type of Amazon EC2 instance. An example is m4.xlarge.

LINKED_ACCOUNT - The description in the attribute map that includes the full name of the member account. The value field contains the Amazon Web Services ID of the member account.

PLATFORM - The Amazon EC2 operating system. Examples are Windows or Linux.

REGION - The Amazon Web Services Region.

SCOPE (Utilization only) - The scope of a Reserved Instance (RI). Values are regional or a single Availability Zone.

TAG (Coverage only) - The tags that are associated with a Reserved Instance (RI).

TENANCY - The tenancy of a resource. Examples are shared or dedicated.

If you set the context to SAVINGS_PLANS, you can use the following dimensions for searching:

SAVINGS_PLANS_TYPE - Type of Savings Plans (EC2 Instance or Compute)

PAYMENT_OPTION - Payment option for the given Savings Plans (for example, All Upfront)

REGION - The Amazon Web Services Region.

INSTANCE_TYPE_FAMILY - The family of instances (For example, m5)

LINKED_ACCOUNT - The description in the attribute map that includes the full name of the member account. The value field contains the Amazon Web Services ID of the member account.

SAVINGS_PLAN_ARN - The unique identifier for your Savings Plan

returnSize :: Int

The number of results that Amazon Web Services returned at one time.

totalSize :: Int

The total number of search results.

newGetDimensionValuesResponse

Arguments

:: Int
$sel:httpStatus:GetDimensionValuesResponse'

-> Int
$sel:returnSize:GetDimensionValuesResponse'

-> Int
$sel:totalSize:GetDimensionValuesResponse'

-> GetDimensionValuesResponse

Create a value of GetDimensionValuesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextPageToken:GetDimensionValues', getDimensionValuesResponse_nextPageToken - The token for the next set of retrievable results. Amazon Web Services provides the token when the response from a previous call has more results than the maximum page size.

$sel:httpStatus:GetDimensionValuesResponse', getDimensionValuesResponse_httpStatus - The response's http status code.

$sel:dimensionValues:GetDimensionValuesResponse', getDimensionValuesResponse_dimensionValues - The filters that you used to filter your request. Some dimensions are available only for a specific context.

If you set the context to COST_AND_USAGE, you can use the following dimensions for searching:

AZ - The Availability Zone. An example is us-east-1a.

DATABASE_ENGINE - The Amazon Relational Database Service database. Examples are Aurora or MySQL.

INSTANCE_TYPE - The type of Amazon EC2 instance. An example is m4.xlarge.

LEGAL_ENTITY_NAME - The name of the organization that sells you Amazon Web Services services, such as Amazon Web Services.

LINKED_ACCOUNT - The description in the attribute map that includes the full name of the member account. The value field contains the Amazon Web Services ID of the member account.

OPERATING_SYSTEM - The operating system. Examples are Windows or Linux.

OPERATION - The action performed. Examples include RunInstance and CreateBucket.

PLATFORM - The Amazon EC2 operating system. Examples are Windows or Linux.

PURCHASE_TYPE - The reservation type of the purchase to which this usage is related. Examples include On-Demand Instances and Standard Reserved Instances.

SERVICE - The Amazon Web Services service such as Amazon DynamoDB.

USAGE_TYPE - The type of usage. An example is DataTransfer-In-Bytes. The response for the GetDimensionValues operation includes a unit attribute. Examples include GB and Hrs.

USAGE_TYPE_GROUP - The grouping of common usage types. An example is Amazon EC2: CloudWatch – Alarms. The response for this operation includes a unit attribute.

RECORD_TYPE - The different types of charges such as RI fees, usage costs, tax refunds, and credits.

RESOURCE_ID - The unique identifier of the resource. ResourceId is an opt-in feature only available for last 14 days for EC2-Compute Service.

If you set the context to RESERVATIONS, you can use the following dimensions for searching:

AZ - The Availability Zone. An example is us-east-1a.

CACHE_ENGINE - The Amazon ElastiCache operating system. Examples are Windows or Linux.

DEPLOYMENT_OPTION - The scope of Amazon Relational Database Service deployments. Valid values are SingleAZ and MultiAZ.

INSTANCE_TYPE - The type of Amazon EC2 instance. An example is m4.xlarge.

LINKED_ACCOUNT - The description in the attribute map that includes the full name of the member account. The value field contains the Amazon Web Services ID of the member account.

PLATFORM - The Amazon EC2 operating system. Examples are Windows or Linux.

REGION - The Amazon Web Services Region.

SCOPE (Utilization only) - The scope of a Reserved Instance (RI). Values are regional or a single Availability Zone.

TAG (Coverage only) - The tags that are associated with a Reserved Instance (RI).

TENANCY - The tenancy of a resource. Examples are shared or dedicated.

If you set the context to SAVINGS_PLANS, you can use the following dimensions for searching:

SAVINGS_PLANS_TYPE - Type of Savings Plans (EC2 Instance or Compute)

PAYMENT_OPTION - Payment option for the given Savings Plans (for example, All Upfront)

REGION - The Amazon Web Services Region.

INSTANCE_TYPE_FAMILY - The family of instances (For example, m5)

LINKED_ACCOUNT - The description in the attribute map that includes the full name of the member account. The value field contains the Amazon Web Services ID of the member account.

SAVINGS_PLAN_ARN - The unique identifier for your Savings Plan

$sel:returnSize:GetDimensionValuesResponse', getDimensionValuesResponse_returnSize - The number of results that Amazon Web Services returned at one time.

$sel:totalSize:GetDimensionValuesResponse', getDimensionValuesResponse_totalSize - The total number of search results.

Response Lenses

getDimensionValuesResponse_nextPageToken :: Lens' GetDimensionValuesResponse (Maybe Text)

The token for the next set of retrievable results. Amazon Web Services provides the token when the response from a previous call has more results than the maximum page size.

getDimensionValuesResponse_httpStatus :: Lens' GetDimensionValuesResponse Int

The response's http status code.

getDimensionValuesResponse_dimensionValues :: Lens' GetDimensionValuesResponse [DimensionValuesWithAttributes]

The filters that you used to filter your request. Some dimensions are available only for a specific context.

If you set the context to COST_AND_USAGE, you can use the following dimensions for searching:

AZ - The Availability Zone. An example is us-east-1a.

DATABASE_ENGINE - The Amazon Relational Database Service database. Examples are Aurora or MySQL.

INSTANCE_TYPE - The type of Amazon EC2 instance. An example is m4.xlarge.

LEGAL_ENTITY_NAME - The name of the organization that sells you Amazon Web Services services, such as Amazon Web Services.

LINKED_ACCOUNT - The description in the attribute map that includes the full name of the member account. The value field contains the Amazon Web Services ID of the member account.

OPERATING_SYSTEM - The operating system. Examples are Windows or Linux.

OPERATION - The action performed. Examples include RunInstance and CreateBucket.

PLATFORM - The Amazon EC2 operating system. Examples are Windows or Linux.

PURCHASE_TYPE - The reservation type of the purchase to which this usage is related. Examples include On-Demand Instances and Standard Reserved Instances.

SERVICE - The Amazon Web Services service such as Amazon DynamoDB.

USAGE_TYPE - The type of usage. An example is DataTransfer-In-Bytes. The response for the GetDimensionValues operation includes a unit attribute. Examples include GB and Hrs.

USAGE_TYPE_GROUP - The grouping of common usage types. An example is Amazon EC2: CloudWatch – Alarms. The response for this operation includes a unit attribute.

RECORD_TYPE - The different types of charges such as RI fees, usage costs, tax refunds, and credits.

RESOURCE_ID - The unique identifier of the resource. ResourceId is an opt-in feature only available for last 14 days for EC2-Compute Service.

If you set the context to RESERVATIONS, you can use the following dimensions for searching:

AZ - The Availability Zone. An example is us-east-1a.

CACHE_ENGINE - The Amazon ElastiCache operating system. Examples are Windows or Linux.

DEPLOYMENT_OPTION - The scope of Amazon Relational Database Service deployments. Valid values are SingleAZ and MultiAZ.

INSTANCE_TYPE - The type of Amazon EC2 instance. An example is m4.xlarge.

LINKED_ACCOUNT - The description in the attribute map that includes the full name of the member account. The value field contains the Amazon Web Services ID of the member account.

PLATFORM - The Amazon EC2 operating system. Examples are Windows or Linux.

REGION - The Amazon Web Services Region.

SCOPE (Utilization only) - The scope of a Reserved Instance (RI). Values are regional or a single Availability Zone.

TAG (Coverage only) - The tags that are associated with a Reserved Instance (RI).

TENANCY - The tenancy of a resource. Examples are shared or dedicated.

If you set the context to SAVINGS_PLANS, you can use the following dimensions for searching:

SAVINGS_PLANS_TYPE - Type of Savings Plans (EC2 Instance or Compute)

PAYMENT_OPTION - Payment option for the given Savings Plans (for example, All Upfront)

REGION - The Amazon Web Services Region.

INSTANCE_TYPE_FAMILY - The family of instances (For example, m5)

LINKED_ACCOUNT - The description in the attribute map that includes the full name of the member account. The value field contains the Amazon Web Services ID of the member account.

SAVINGS_PLAN_ARN - The unique identifier for your Savings Plan

getDimensionValuesResponse_returnSize :: Lens' GetDimensionValuesResponse Int

The number of results that Amazon Web Services returned at one time.

getDimensionValuesResponse_totalSize :: Lens' GetDimensionValuesResponse Int

The total number of search results.