Network.AWS.CostExplorer.GetReservationCoverage

Description

Retrieves the reservation coverage for your account. This enables you to see how much of your Amazon Elastic Compute Cloud, Amazon ElastiCache, Amazon Relational Database Service, or Amazon Redshift usage is covered by a reservation. An organization's management account can see the coverage of the associated member accounts. This supports dimensions, Cost Categories, and nested expressions. For any time period, you can filter data about reservation usage by the following dimensions:

AZ

CACHE_ENGINE

DATABASE_ENGINE

DEPLOYMENT_OPTION

INSTANCE_TYPE

LINKED_ACCOUNT

OPERATING_SYSTEM

PLATFORM

REGION

SERVICE

TAG

TENANCY

To determine valid values for a dimension, use the GetDimensionValues operation.

Creating a Request

data GetReservationCoverage

You can use the following request parameters to query for how much of your instance usage a reservation covered.

See: newGetReservationCoverage smart constructor.

Constructors

GetReservationCoverage'

Fields

groupBy :: Maybe [GroupDefinition]

You can group the data by the following attributes:

AZ

CACHE_ENGINE

DATABASE_ENGINE

DEPLOYMENT_OPTION

INSTANCE_TYPE

LINKED_ACCOUNT

OPERATING_SYSTEM

PLATFORM

REGION

TENANCY

nextPageToken :: Maybe Text

The token to retrieve the next set of results. Amazon Web Services provides the token when the response from a previous call has more results than the maximum page size.

metrics :: Maybe [Text]

The measurement that you want your reservation coverage reported in.

Valid values are Hour, Unit, and Cost. You can use multiple values in a request.

granularity :: Maybe Granularity

The granularity of the Amazon Web Services cost data for the reservation. Valid values are MONTHLY and DAILY.

If GroupBy is set, Granularity can't be set. If Granularity isn't set, the response object doesn't include Granularity, either MONTHLY or DAILY.

The GetReservationCoverage operation supports only DAILY and MONTHLY granularities.

filter' :: Maybe Expression

Filters utilization data by dimensions. You can filter by the following dimensions:

AZ

CACHE_ENGINE

DATABASE_ENGINE

DEPLOYMENT_OPTION

INSTANCE_TYPE

LINKED_ACCOUNT

OPERATING_SYSTEM

PLATFORM

REGION

SERVICE

TAG

TENANCY

GetReservationCoverage uses the same Expression object as the other operations, but only AND is supported among each dimension. You can nest only one level deep. If there are multiple values for a dimension, they are OR'd together.

If you don't provide a SERVICE filter, Cost Explorer defaults to EC2.

Cost category is also supported.

maxResults :: Maybe Natural

The maximum number of objects that you returned for this request. If more objects are available, in the response, Amazon Web Services provides a NextPageToken value that you can use in a subsequent call to get the next batch of objects.

sortBy :: Maybe SortDefinition

The value by which you want to sort the data.

The following values are supported for Key:

OnDemandCost

CoverageHoursPercentage

OnDemandHours

ReservedHours

TotalRunningHours

CoverageNormalizedUnitsPercentage

OnDemandNormalizedUnits

ReservedNormalizedUnits

TotalRunningNormalizedUnits

Time

Supported values for SortOrder are ASCENDING or DESCENDING.

timePeriod :: DateInterval

The start and end dates of the period that you want to retrieve data about reservation coverage for. You can retrieve data for a maximum of 13 months: the last 12 months and the current month. The start date is inclusive, but the end date is exclusive. For example, if start is 2017-01-01 and end is 2017-05-01, then the cost and usage data is retrieved from 2017-01-01 up to and including 2017-04-30 but not including 2017-05-01.

newGetReservationCoverage

Arguments

:: DateInterval
$sel:timePeriod:GetReservationCoverage'

-> GetReservationCoverage

Create a value of GetReservationCoverage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groupBy:GetReservationCoverage', getReservationCoverage_groupBy - You can group the data by the following attributes:

AZ

CACHE_ENGINE

DATABASE_ENGINE

DEPLOYMENT_OPTION

INSTANCE_TYPE

LINKED_ACCOUNT

OPERATING_SYSTEM

PLATFORM

REGION

TENANCY

$sel:nextPageToken:GetReservationCoverage', getReservationCoverage_nextPageToken - The token to retrieve the next set of results. Amazon Web Services provides the token when the response from a previous call has more results than the maximum page size.

$sel:metrics:GetReservationCoverage', getReservationCoverage_metrics - The measurement that you want your reservation coverage reported in.

Valid values are Hour, Unit, and Cost. You can use multiple values in a request.

$sel:granularity:GetReservationCoverage', getReservationCoverage_granularity - The granularity of the Amazon Web Services cost data for the reservation. Valid values are MONTHLY and DAILY.

If GroupBy is set, Granularity can't be set. If Granularity isn't set, the response object doesn't include Granularity, either MONTHLY or DAILY.

The GetReservationCoverage operation supports only DAILY and MONTHLY granularities.

$sel:filter':GetReservationCoverage', getReservationCoverage_filter - Filters utilization data by dimensions. You can filter by the following dimensions:

AZ

CACHE_ENGINE

DATABASE_ENGINE

DEPLOYMENT_OPTION

INSTANCE_TYPE

LINKED_ACCOUNT

OPERATING_SYSTEM

PLATFORM

REGION

SERVICE

TAG

TENANCY

GetReservationCoverage uses the same Expression object as the other operations, but only AND is supported among each dimension. You can nest only one level deep. If there are multiple values for a dimension, they are OR'd together.

If you don't provide a SERVICE filter, Cost Explorer defaults to EC2.

Cost category is also supported.

$sel:maxResults:GetReservationCoverage', getReservationCoverage_maxResults - The maximum number of objects that you returned for this request. If more objects are available, in the response, Amazon Web Services provides a NextPageToken value that you can use in a subsequent call to get the next batch of objects.

$sel:sortBy:GetReservationCoverage', getReservationCoverage_sortBy - The value by which you want to sort the data.

The following values are supported for Key:

OnDemandCost

CoverageHoursPercentage

OnDemandHours

ReservedHours

TotalRunningHours

CoverageNormalizedUnitsPercentage

OnDemandNormalizedUnits

ReservedNormalizedUnits

TotalRunningNormalizedUnits

Time

Supported values for SortOrder are ASCENDING or DESCENDING.

$sel:timePeriod:GetReservationCoverage', getReservationCoverage_timePeriod - The start and end dates of the period that you want to retrieve data about reservation coverage for. You can retrieve data for a maximum of 13 months: the last 12 months and the current month. The start date is inclusive, but the end date is exclusive. For example, if start is 2017-01-01 and end is 2017-05-01, then the cost and usage data is retrieved from 2017-01-01 up to and including 2017-04-30 but not including 2017-05-01.

Request Lenses

getReservationCoverage_groupBy :: Lens' GetReservationCoverage (Maybe [GroupDefinition])

You can group the data by the following attributes:

AZ

CACHE_ENGINE

DATABASE_ENGINE

DEPLOYMENT_OPTION

INSTANCE_TYPE

LINKED_ACCOUNT

OPERATING_SYSTEM

PLATFORM

REGION

TENANCY

getReservationCoverage_nextPageToken :: Lens' GetReservationCoverage (Maybe Text)

The token to retrieve the next set of results. Amazon Web Services provides the token when the response from a previous call has more results than the maximum page size.

getReservationCoverage_metrics :: Lens' GetReservationCoverage (Maybe [Text])

The measurement that you want your reservation coverage reported in.

Valid values are Hour, Unit, and Cost. You can use multiple values in a request.

getReservationCoverage_granularity :: Lens' GetReservationCoverage (Maybe Granularity)

The granularity of the Amazon Web Services cost data for the reservation. Valid values are MONTHLY and DAILY.

If GroupBy is set, Granularity can't be set. If Granularity isn't set, the response object doesn't include Granularity, either MONTHLY or DAILY.

The GetReservationCoverage operation supports only DAILY and MONTHLY granularities.

getReservationCoverage_filter :: Lens' GetReservationCoverage (Maybe Expression)

Filters utilization data by dimensions. You can filter by the following dimensions:

AZ

CACHE_ENGINE

DATABASE_ENGINE

DEPLOYMENT_OPTION

INSTANCE_TYPE

LINKED_ACCOUNT

OPERATING_SYSTEM

PLATFORM

REGION

SERVICE

TAG

TENANCY

GetReservationCoverage uses the same Expression object as the other operations, but only AND is supported among each dimension. You can nest only one level deep. If there are multiple values for a dimension, they are OR'd together.

If you don't provide a SERVICE filter, Cost Explorer defaults to EC2.

Cost category is also supported.

getReservationCoverage_maxResults :: Lens' GetReservationCoverage (Maybe Natural)

The maximum number of objects that you returned for this request. If more objects are available, in the response, Amazon Web Services provides a NextPageToken value that you can use in a subsequent call to get the next batch of objects.

getReservationCoverage_sortBy :: Lens' GetReservationCoverage (Maybe SortDefinition)

The value by which you want to sort the data.

The following values are supported for Key:

OnDemandCost

CoverageHoursPercentage

OnDemandHours

ReservedHours

TotalRunningHours

CoverageNormalizedUnitsPercentage

OnDemandNormalizedUnits

ReservedNormalizedUnits

TotalRunningNormalizedUnits

Time

Supported values for SortOrder are ASCENDING or DESCENDING.

getReservationCoverage_timePeriod :: Lens' GetReservationCoverage DateInterval

The start and end dates of the period that you want to retrieve data about reservation coverage for. You can retrieve data for a maximum of 13 months: the last 12 months and the current month. The start date is inclusive, but the end date is exclusive. For example, if start is 2017-01-01 and end is 2017-05-01, then the cost and usage data is retrieved from 2017-01-01 up to and including 2017-04-30 but not including 2017-05-01.

Destructuring the Response

data GetReservationCoverageResponse

See: newGetReservationCoverageResponse smart constructor.

Constructors

GetReservationCoverageResponse'

Fields

nextPageToken :: Maybe Text

The token for the next set of retrievable results. Amazon Web Services provides the token when the response from a previous call has more results than the maximum page size.

total :: Maybe Coverage

The total amount of instance usage that a reservation covered.

httpStatus :: Int

The response's http status code.

coveragesByTime :: [CoverageByTime]

The amount of time that your reservations covered.

newGetReservationCoverageResponse

Arguments

:: Int
$sel:httpStatus:GetReservationCoverageResponse'

-> GetReservationCoverageResponse

Create a value of GetReservationCoverageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextPageToken:GetReservationCoverage', getReservationCoverageResponse_nextPageToken - The token for the next set of retrievable results. Amazon Web Services provides the token when the response from a previous call has more results than the maximum page size.

$sel:total:GetReservationCoverageResponse', getReservationCoverageResponse_total - The total amount of instance usage that a reservation covered.

$sel:httpStatus:GetReservationCoverageResponse', getReservationCoverageResponse_httpStatus - The response's http status code.

$sel:coveragesByTime:GetReservationCoverageResponse', getReservationCoverageResponse_coveragesByTime - The amount of time that your reservations covered.

Response Lenses

getReservationCoverageResponse_nextPageToken :: Lens' GetReservationCoverageResponse (Maybe Text)

The token for the next set of retrievable results. Amazon Web Services provides the token when the response from a previous call has more results than the maximum page size.

getReservationCoverageResponse_total :: Lens' GetReservationCoverageResponse (Maybe Coverage)

The total amount of instance usage that a reservation covered.

getReservationCoverageResponse_httpStatus :: Lens' GetReservationCoverageResponse Int

The response's http status code.

getReservationCoverageResponse_coveragesByTime :: Lens' GetReservationCoverageResponse [CoverageByTime]

The amount of time that your reservations covered.