Network.AWS.CostExplorer.GetRightsizingRecommendation

Description

Creates recommendations that help you save cost by identifying idle and underutilized Amazon EC2 instances.

Recommendations are generated to either downsize or terminate instances, along with providing savings detail and metrics. For details on calculation and function, see Optimizing Your Cost with Rightsizing Recommendations in the Billing and Cost Management User Guide.

Creating a Request

data GetRightsizingRecommendation

See: newGetRightsizingRecommendation smart constructor.

Constructors

GetRightsizingRecommendation'

Fields

pageSize :: Maybe Natural

The number of recommendations that you want returned in a single response object.

configuration :: Maybe RightsizingRecommendationConfiguration

Enables you to customize recommendations across two attributes. You can choose to view recommendations for instances within the same instance families or across different instance families. You can also choose to view your estimated savings associated with recommendations with consideration of existing Savings Plans or RI benefits, or neither.

nextPageToken :: Maybe Text

The pagination token that indicates the next set of results that you want to retrieve.

filter' :: Maybe Expression

service :: Text

The specific service that you want recommendations for. The only valid value for GetRightsizingRecommendation is "AmazonEC2".

newGetRightsizingRecommendation

Arguments

:: Text
$sel:service:GetRightsizingRecommendation'

-> GetRightsizingRecommendation

Create a value of GetRightsizingRecommendation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pageSize:GetRightsizingRecommendation', getRightsizingRecommendation_pageSize - The number of recommendations that you want returned in a single response object.

$sel:configuration:GetRightsizingRecommendation', getRightsizingRecommendation_configuration - Enables you to customize recommendations across two attributes. You can choose to view recommendations for instances within the same instance families or across different instance families. You can also choose to view your estimated savings associated with recommendations with consideration of existing Savings Plans or RI benefits, or neither.

$sel:nextPageToken:GetRightsizingRecommendation', getRightsizingRecommendation_nextPageToken - The pagination token that indicates the next set of results that you want to retrieve.

$sel:filter':GetRightsizingRecommendation', getRightsizingRecommendation_filter - Undocumented member.

$sel:service:GetRightsizingRecommendation', getRightsizingRecommendation_service - The specific service that you want recommendations for. The only valid value for GetRightsizingRecommendation is "AmazonEC2".

Request Lenses

getRightsizingRecommendation_pageSize :: Lens' GetRightsizingRecommendation (Maybe Natural)

The number of recommendations that you want returned in a single response object.

getRightsizingRecommendation_configuration :: Lens' GetRightsizingRecommendation (Maybe RightsizingRecommendationConfiguration)

Enables you to customize recommendations across two attributes. You can choose to view recommendations for instances within the same instance families or across different instance families. You can also choose to view your estimated savings associated with recommendations with consideration of existing Savings Plans or RI benefits, or neither.

getRightsizingRecommendation_nextPageToken :: Lens' GetRightsizingRecommendation (Maybe Text)

The pagination token that indicates the next set of results that you want to retrieve.

getRightsizingRecommendation_filter :: Lens' GetRightsizingRecommendation (Maybe Expression)

Undocumented member.

getRightsizingRecommendation_service :: Lens' GetRightsizingRecommendation Text

The specific service that you want recommendations for. The only valid value for GetRightsizingRecommendation is "AmazonEC2".

Destructuring the Response

data GetRightsizingRecommendationResponse

See: newGetRightsizingRecommendationResponse smart constructor.

Constructors

GetRightsizingRecommendationResponse'

Fields

configuration :: Maybe RightsizingRecommendationConfiguration

Enables you to customize recommendations across two attributes. You can choose to view recommendations for instances within the same instance families or across different instance families. You can also choose to view your estimated savings associated with recommendations with consideration of existing Savings Plans or RI benefits, or neither.

metadata :: Maybe RightsizingRecommendationMetadata

Information regarding this specific recommendation set.

nextPageToken :: Maybe Text

The token to retrieve the next set of results.

summary :: Maybe RightsizingRecommendationSummary

Summary of this recommendation set.

rightsizingRecommendations :: Maybe [RightsizingRecommendation]

Recommendations to rightsize resources.

httpStatus :: Int

The response's http status code.

newGetRightsizingRecommendationResponse

Arguments

:: Int
$sel:httpStatus:GetRightsizingRecommendationResponse'

-> GetRightsizingRecommendationResponse

Create a value of GetRightsizingRecommendationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configuration:GetRightsizingRecommendation', getRightsizingRecommendationResponse_configuration - Enables you to customize recommendations across two attributes. You can choose to view recommendations for instances within the same instance families or across different instance families. You can also choose to view your estimated savings associated with recommendations with consideration of existing Savings Plans or RI benefits, or neither.

$sel:metadata:GetRightsizingRecommendationResponse', getRightsizingRecommendationResponse_metadata - Information regarding this specific recommendation set.

$sel:nextPageToken:GetRightsizingRecommendation', getRightsizingRecommendationResponse_nextPageToken - The token to retrieve the next set of results.

$sel:summary:GetRightsizingRecommendationResponse', getRightsizingRecommendationResponse_summary - Summary of this recommendation set.

$sel:rightsizingRecommendations:GetRightsizingRecommendationResponse', getRightsizingRecommendationResponse_rightsizingRecommendations - Recommendations to rightsize resources.

$sel:httpStatus:GetRightsizingRecommendationResponse', getRightsizingRecommendationResponse_httpStatus - The response's http status code.

Response Lenses

getRightsizingRecommendationResponse_configuration :: Lens' GetRightsizingRecommendationResponse (Maybe RightsizingRecommendationConfiguration)

Enables you to customize recommendations across two attributes. You can choose to view recommendations for instances within the same instance families or across different instance families. You can also choose to view your estimated savings associated with recommendations with consideration of existing Savings Plans or RI benefits, or neither.

getRightsizingRecommendationResponse_metadata :: Lens' GetRightsizingRecommendationResponse (Maybe RightsizingRecommendationMetadata)

Information regarding this specific recommendation set.

getRightsizingRecommendationResponse_nextPageToken :: Lens' GetRightsizingRecommendationResponse (Maybe Text)

The token to retrieve the next set of results.

getRightsizingRecommendationResponse_summary :: Lens' GetRightsizingRecommendationResponse (Maybe RightsizingRecommendationSummary)

Summary of this recommendation set.

getRightsizingRecommendationResponse_rightsizingRecommendations :: Lens' GetRightsizingRecommendationResponse (Maybe [RightsizingRecommendation])

Recommendations to rightsize resources.

getRightsizingRecommendationResponse_httpStatus :: Lens' GetRightsizingRecommendationResponse Int

The response's http status code.