Network.AWS.CostExplorer.GetSavingsPlansPurchaseRecommendation

Description

Retrieves your request parameters, Savings Plan Recommendations Summary and Details.

Creating a Request

data GetSavingsPlansPurchaseRecommendation

See: newGetSavingsPlansPurchaseRecommendation smart constructor.

Constructors

GetSavingsPlansPurchaseRecommendation'

Fields

pageSize :: Maybe Natural

The number of recommendations that you want returned in a single response object.

accountScope :: Maybe AccountScope

The account scope that you want your recommendations for. Amazon Web Services calculates recommendations including the management account and member accounts if the value is set to PAYER. If the value is LINKED, recommendations are calculated for individual member accounts only.

nextPageToken :: Maybe Text

The token to retrieve the next set of results. Amazon Web Services provides the token when the response from a previous call has more results than the maximum page size.

filter' :: Maybe Expression

You can filter your recommendations by Account ID with the LINKED_ACCOUNT dimension. To filter your recommendations by Account ID, specify Key as LINKED_ACCOUNT and Value as the comma-separated Acount ID(s) for which you want to see Savings Plans purchase recommendations.

For GetSavingsPlansPurchaseRecommendation, the Filter does not include CostCategories or Tags. It only includes Dimensions. With Dimensions, Key must be LINKED_ACCOUNT and Value can be a single Account ID or multiple comma-separated Account IDs for which you want to see Savings Plans Purchase Recommendations. AND and OR operators are not supported.

savingsPlansType :: SupportedSavingsPlansType

The Savings Plans recommendation type requested.

termInYears :: TermInYears

The savings plan recommendation term used to generate these recommendations.

paymentOption :: PaymentOption

The payment option used to generate these recommendations.

lookbackPeriodInDays :: LookbackPeriodInDays

The lookback period used to generate the recommendation.

newGetSavingsPlansPurchaseRecommendation

Arguments

:: SupportedSavingsPlansType
$sel:savingsPlansType:GetSavingsPlansPurchaseRecommendation'

-> TermInYears
$sel:termInYears:GetSavingsPlansPurchaseRecommendation'

-> PaymentOption
$sel:paymentOption:GetSavingsPlansPurchaseRecommendation'

-> LookbackPeriodInDays
$sel:lookbackPeriodInDays:GetSavingsPlansPurchaseRecommendation'

-> GetSavingsPlansPurchaseRecommendation

Create a value of GetSavingsPlansPurchaseRecommendation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pageSize:GetSavingsPlansPurchaseRecommendation', getSavingsPlansPurchaseRecommendation_pageSize - The number of recommendations that you want returned in a single response object.

$sel:accountScope:GetSavingsPlansPurchaseRecommendation', getSavingsPlansPurchaseRecommendation_accountScope - The account scope that you want your recommendations for. Amazon Web Services calculates recommendations including the management account and member accounts if the value is set to PAYER. If the value is LINKED, recommendations are calculated for individual member accounts only.

$sel:nextPageToken:GetSavingsPlansPurchaseRecommendation', getSavingsPlansPurchaseRecommendation_nextPageToken - The token to retrieve the next set of results. Amazon Web Services provides the token when the response from a previous call has more results than the maximum page size.

$sel:filter':GetSavingsPlansPurchaseRecommendation', getSavingsPlansPurchaseRecommendation_filter - You can filter your recommendations by Account ID with the LINKED_ACCOUNT dimension. To filter your recommendations by Account ID, specify Key as LINKED_ACCOUNT and Value as the comma-separated Acount ID(s) for which you want to see Savings Plans purchase recommendations.

For GetSavingsPlansPurchaseRecommendation, the Filter does not include CostCategories or Tags. It only includes Dimensions. With Dimensions, Key must be LINKED_ACCOUNT and Value can be a single Account ID or multiple comma-separated Account IDs for which you want to see Savings Plans Purchase Recommendations. AND and OR operators are not supported.

$sel:savingsPlansType:GetSavingsPlansPurchaseRecommendation', getSavingsPlansPurchaseRecommendation_savingsPlansType - The Savings Plans recommendation type requested.

$sel:termInYears:GetSavingsPlansPurchaseRecommendation', getSavingsPlansPurchaseRecommendation_termInYears - The savings plan recommendation term used to generate these recommendations.

$sel:paymentOption:GetSavingsPlansPurchaseRecommendation', getSavingsPlansPurchaseRecommendation_paymentOption - The payment option used to generate these recommendations.

$sel:lookbackPeriodInDays:GetSavingsPlansPurchaseRecommendation', getSavingsPlansPurchaseRecommendation_lookbackPeriodInDays - The lookback period used to generate the recommendation.

Request Lenses

getSavingsPlansPurchaseRecommendation_pageSize :: Lens' GetSavingsPlansPurchaseRecommendation (Maybe Natural)

The number of recommendations that you want returned in a single response object.

getSavingsPlansPurchaseRecommendation_accountScope :: Lens' GetSavingsPlansPurchaseRecommendation (Maybe AccountScope)

The account scope that you want your recommendations for. Amazon Web Services calculates recommendations including the management account and member accounts if the value is set to PAYER. If the value is LINKED, recommendations are calculated for individual member accounts only.

getSavingsPlansPurchaseRecommendation_nextPageToken :: Lens' GetSavingsPlansPurchaseRecommendation (Maybe Text)

The token to retrieve the next set of results. Amazon Web Services provides the token when the response from a previous call has more results than the maximum page size.

getSavingsPlansPurchaseRecommendation_filter :: Lens' GetSavingsPlansPurchaseRecommendation (Maybe Expression)

You can filter your recommendations by Account ID with the LINKED_ACCOUNT dimension. To filter your recommendations by Account ID, specify Key as LINKED_ACCOUNT and Value as the comma-separated Acount ID(s) for which you want to see Savings Plans purchase recommendations.

For GetSavingsPlansPurchaseRecommendation, the Filter does not include CostCategories or Tags. It only includes Dimensions. With Dimensions, Key must be LINKED_ACCOUNT and Value can be a single Account ID or multiple comma-separated Account IDs for which you want to see Savings Plans Purchase Recommendations. AND and OR operators are not supported.

getSavingsPlansPurchaseRecommendation_savingsPlansType :: Lens' GetSavingsPlansPurchaseRecommendation SupportedSavingsPlansType

The Savings Plans recommendation type requested.

getSavingsPlansPurchaseRecommendation_termInYears :: Lens' GetSavingsPlansPurchaseRecommendation TermInYears

The savings plan recommendation term used to generate these recommendations.

getSavingsPlansPurchaseRecommendation_paymentOption :: Lens' GetSavingsPlansPurchaseRecommendation PaymentOption

The payment option used to generate these recommendations.

getSavingsPlansPurchaseRecommendation_lookbackPeriodInDays :: Lens' GetSavingsPlansPurchaseRecommendation LookbackPeriodInDays

The lookback period used to generate the recommendation.

Destructuring the Response

data GetSavingsPlansPurchaseRecommendationResponse

See: newGetSavingsPlansPurchaseRecommendationResponse smart constructor.

Constructors

GetSavingsPlansPurchaseRecommendationResponse'

Fields

metadata :: Maybe SavingsPlansPurchaseRecommendationMetadata

Information regarding this specific recommendation set.

nextPageToken :: Maybe Text

The token for the next set of retrievable results. Amazon Web Services provides the token when the response from a previous call has more results than the maximum page size.

savingsPlansPurchaseRecommendation :: Maybe SavingsPlansPurchaseRecommendation

Contains your request parameters, Savings Plan Recommendations Summary, and Details.

httpStatus :: Int

The response's http status code.

newGetSavingsPlansPurchaseRecommendationResponse

Arguments

:: Int
$sel:httpStatus:GetSavingsPlansPurchaseRecommendationResponse'

-> GetSavingsPlansPurchaseRecommendationResponse

Create a value of GetSavingsPlansPurchaseRecommendationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metadata:GetSavingsPlansPurchaseRecommendationResponse', getSavingsPlansPurchaseRecommendationResponse_metadata - Information regarding this specific recommendation set.

$sel:nextPageToken:GetSavingsPlansPurchaseRecommendation', getSavingsPlansPurchaseRecommendationResponse_nextPageToken - The token for the next set of retrievable results. Amazon Web Services provides the token when the response from a previous call has more results than the maximum page size.

$sel:savingsPlansPurchaseRecommendation:GetSavingsPlansPurchaseRecommendationResponse', getSavingsPlansPurchaseRecommendationResponse_savingsPlansPurchaseRecommendation - Contains your request parameters, Savings Plan Recommendations Summary, and Details.

$sel:httpStatus:GetSavingsPlansPurchaseRecommendationResponse', getSavingsPlansPurchaseRecommendationResponse_httpStatus - The response's http status code.

Response Lenses

getSavingsPlansPurchaseRecommendationResponse_metadata :: Lens' GetSavingsPlansPurchaseRecommendationResponse (Maybe SavingsPlansPurchaseRecommendationMetadata)

Information regarding this specific recommendation set.

getSavingsPlansPurchaseRecommendationResponse_nextPageToken :: Lens' GetSavingsPlansPurchaseRecommendationResponse (Maybe Text)

The token for the next set of retrievable results. Amazon Web Services provides the token when the response from a previous call has more results than the maximum page size.

getSavingsPlansPurchaseRecommendationResponse_savingsPlansPurchaseRecommendation :: Lens' GetSavingsPlansPurchaseRecommendationResponse (Maybe SavingsPlansPurchaseRecommendation)

Contains your request parameters, Savings Plan Recommendations Summary, and Details.

getSavingsPlansPurchaseRecommendationResponse_httpStatus :: Lens' GetSavingsPlansPurchaseRecommendationResponse Int

The response's http status code.