Network.AWS.CostExplorer.GetTags

Description

Queries for available tag keys and tag values for a specified period. You can search the tag values for an arbitrary string.

Creating a Request

data GetTags

See: newGetTags smart constructor.

Constructors

GetTags'

Fields

nextPageToken :: Maybe Text

The token to retrieve the next set of results. Amazon Web Services provides the token when the response from a previous call has more results than the maximum page size.

searchString :: Maybe Text

The value that you want to search for.

tagKey :: Maybe Text

The key of the tag that you want to return values for.

filter' :: Maybe Expression

maxResults :: Maybe Natural

This field is only used when SortBy is provided in the request. The maximum number of objects that to be returned for this request. If MaxResults is not specified with SortBy, the request will return 1000 results as the default value for this parameter.

For GetTags, MaxResults has an upper limit of 1000.

sortBy :: Maybe [SortDefinition]

The value by which you want to sort the data.

The key represents cost and usage metrics. The following values are supported:

BlendedCost

UnblendedCost

AmortizedCost

NetAmortizedCost

NetUnblendedCost

UsageQuantity

NormalizedUsageAmount

Supported values for SortOrder are ASCENDING or DESCENDING.

When using SortBy, NextPageToken and SearchString are not supported.

timePeriod :: DateInterval

The start and end dates for retrieving the dimension values. The start date is inclusive, but the end date is exclusive. For example, if start is 2017-01-01 and end is 2017-05-01, then the cost and usage data is retrieved from 2017-01-01 up to and including 2017-04-30 but not including 2017-05-01.

newGetTags

Arguments

:: DateInterval
$sel:timePeriod:GetTags'

-> GetTags

Create a value of GetTags with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextPageToken:GetTags', getTags_nextPageToken - The token to retrieve the next set of results. Amazon Web Services provides the token when the response from a previous call has more results than the maximum page size.

$sel:searchString:GetTags', getTags_searchString - The value that you want to search for.

$sel:tagKey:GetTags', getTags_tagKey - The key of the tag that you want to return values for.

$sel:filter':GetTags', getTags_filter - Undocumented member.

$sel:maxResults:GetTags', getTags_maxResults - This field is only used when SortBy is provided in the request. The maximum number of objects that to be returned for this request. If MaxResults is not specified with SortBy, the request will return 1000 results as the default value for this parameter.

For GetTags, MaxResults has an upper limit of 1000.

$sel:sortBy:GetTags', getTags_sortBy - The value by which you want to sort the data.

The key represents cost and usage metrics. The following values are supported:

BlendedCost

UnblendedCost

AmortizedCost

NetAmortizedCost

NetUnblendedCost

UsageQuantity

NormalizedUsageAmount

Supported values for SortOrder are ASCENDING or DESCENDING.

When using SortBy, NextPageToken and SearchString are not supported.

$sel:timePeriod:GetTags', getTags_timePeriod - The start and end dates for retrieving the dimension values. The start date is inclusive, but the end date is exclusive. For example, if start is 2017-01-01 and end is 2017-05-01, then the cost and usage data is retrieved from 2017-01-01 up to and including 2017-04-30 but not including 2017-05-01.

Request Lenses

getTags_nextPageToken :: Lens' GetTags (Maybe Text)

The token to retrieve the next set of results. Amazon Web Services provides the token when the response from a previous call has more results than the maximum page size.

getTags_searchString :: Lens' GetTags (Maybe Text)

The value that you want to search for.

getTags_tagKey :: Lens' GetTags (Maybe Text)

The key of the tag that you want to return values for.

getTags_filter :: Lens' GetTags (Maybe Expression)

Undocumented member.

getTags_maxResults :: Lens' GetTags (Maybe Natural)

This field is only used when SortBy is provided in the request. The maximum number of objects that to be returned for this request. If MaxResults is not specified with SortBy, the request will return 1000 results as the default value for this parameter.

For GetTags, MaxResults has an upper limit of 1000.

getTags_sortBy :: Lens' GetTags (Maybe [SortDefinition])

The value by which you want to sort the data.

The key represents cost and usage metrics. The following values are supported:

BlendedCost

UnblendedCost

AmortizedCost

NetAmortizedCost

NetUnblendedCost

UsageQuantity

NormalizedUsageAmount

Supported values for SortOrder are ASCENDING or DESCENDING.

When using SortBy, NextPageToken and SearchString are not supported.

getTags_timePeriod :: Lens' GetTags DateInterval

The start and end dates for retrieving the dimension values. The start date is inclusive, but the end date is exclusive. For example, if start is 2017-01-01 and end is 2017-05-01, then the cost and usage data is retrieved from 2017-01-01 up to and including 2017-04-30 but not including 2017-05-01.

Destructuring the Response

data GetTagsResponse

See: newGetTagsResponse smart constructor.

Constructors

GetTagsResponse'

Fields

nextPageToken :: Maybe Text

The token for the next set of retrievable results. Amazon Web Services provides the token when the response from a previous call has more results than the maximum page size.

httpStatus :: Int

The response's http status code.

tags :: [Text]

The tags that match your request.

returnSize :: Int

The number of query results that Amazon Web Services returns at a time.

totalSize :: Int

The total number of query results.

newGetTagsResponse

Arguments

:: Int
$sel:httpStatus:GetTagsResponse'

-> Int
$sel:returnSize:GetTagsResponse'

-> Int
$sel:totalSize:GetTagsResponse'

-> GetTagsResponse

Create a value of GetTagsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextPageToken:GetTags', getTagsResponse_nextPageToken - The token for the next set of retrievable results. Amazon Web Services provides the token when the response from a previous call has more results than the maximum page size.

$sel:httpStatus:GetTagsResponse', getTagsResponse_httpStatus - The response's http status code.

$sel:tags:GetTagsResponse', getTagsResponse_tags - The tags that match your request.

$sel:returnSize:GetTagsResponse', getTagsResponse_returnSize - The number of query results that Amazon Web Services returns at a time.

$sel:totalSize:GetTagsResponse', getTagsResponse_totalSize - The total number of query results.

Response Lenses

getTagsResponse_nextPageToken :: Lens' GetTagsResponse (Maybe Text)

The token for the next set of retrievable results. Amazon Web Services provides the token when the response from a previous call has more results than the maximum page size.

getTagsResponse_httpStatus :: Lens' GetTagsResponse Int

The response's http status code.

getTagsResponse_tags :: Lens' GetTagsResponse [Text]

The tags that match your request.

getTagsResponse_returnSize :: Lens' GetTagsResponse Int

The number of query results that Amazon Web Services returns at a time.

getTagsResponse_totalSize :: Lens' GetTagsResponse Int

The total number of query results.