Network.AWS.CostExplorer.Types.Anomaly

Description

Documentation

data Anomaly

An unusual cost pattern. This consists of the detailed metadata and the current status of the anomaly object.

See: newAnomaly smart constructor.

Constructors

Anomaly'

Fields

dimensionValue :: Maybe Text

The dimension for the anomaly (for example, an Amazon Web Services service in a service monitor).

rootCauses :: Maybe [RootCause]

The list of identified root causes for the anomaly.

feedback :: Maybe AnomalyFeedbackType

The feedback value.

anomalyStartDate :: Maybe Text

The first day the anomaly is detected.

anomalyEndDate :: Maybe Text

The last day the anomaly is detected.

anomalyId :: Text

The unique identifier for the anomaly.

anomalyScore :: AnomalyScore

The latest and maximum score for the anomaly.

impact :: Impact

The dollar impact for the anomaly.

monitorArn :: Text

The Amazon Resource Name (ARN) for the cost monitor that generated this anomaly.

newAnomaly

Arguments

:: Text
$sel:anomalyId:Anomaly'

-> AnomalyScore
$sel:anomalyScore:Anomaly'

-> Impact
$sel:impact:Anomaly'

-> Text
$sel:monitorArn:Anomaly'

-> Anomaly

Create a value of Anomaly with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dimensionValue:Anomaly', anomaly_dimensionValue - The dimension for the anomaly (for example, an Amazon Web Services service in a service monitor).

$sel:rootCauses:Anomaly', anomaly_rootCauses - The list of identified root causes for the anomaly.

$sel:feedback:Anomaly', anomaly_feedback - The feedback value.

$sel:anomalyStartDate:Anomaly', anomaly_anomalyStartDate - The first day the anomaly is detected.

$sel:anomalyEndDate:Anomaly', anomaly_anomalyEndDate - The last day the anomaly is detected.

$sel:anomalyId:Anomaly', anomaly_anomalyId - The unique identifier for the anomaly.

$sel:anomalyScore:Anomaly', anomaly_anomalyScore - The latest and maximum score for the anomaly.

$sel:impact:Anomaly', anomaly_impact - The dollar impact for the anomaly.

$sel:monitorArn:Anomaly', anomaly_monitorArn - The Amazon Resource Name (ARN) for the cost monitor that generated this anomaly.

anomaly_dimensionValue :: Lens' Anomaly (Maybe Text)

The dimension for the anomaly (for example, an Amazon Web Services service in a service monitor).

anomaly_rootCauses :: Lens' Anomaly (Maybe [RootCause])

The list of identified root causes for the anomaly.

anomaly_feedback :: Lens' Anomaly (Maybe AnomalyFeedbackType)

The feedback value.

anomaly_anomalyStartDate :: Lens' Anomaly (Maybe Text)

The first day the anomaly is detected.

anomaly_anomalyEndDate :: Lens' Anomaly (Maybe Text)

The last day the anomaly is detected.

anomaly_anomalyId :: Lens' Anomaly Text

The unique identifier for the anomaly.

anomaly_anomalyScore :: Lens' Anomaly AnomalyScore

The latest and maximum score for the anomaly.

anomaly_impact :: Lens' Anomaly Impact

The dollar impact for the anomaly.

anomaly_monitorArn :: Lens' Anomaly Text

The Amazon Resource Name (ARN) for the cost monitor that generated this anomaly.