Network.AWS.CostExplorer.Types.AnomalySubscription

Description

Documentation

data AnomalySubscription

The association between a monitor, threshold, and list of subscribers used to deliver notifications about anomalies detected by a monitor that exceeds a threshold. The content consists of the detailed metadata and the current status of the AnomalySubscription object.

See: newAnomalySubscription smart constructor.

Constructors

AnomalySubscription'

Fields

accountId :: Maybe Text

Your unique account identifier.

subscriptionArn :: Maybe Text

The AnomalySubscription Amazon Resource Name (ARN).

monitorArnList :: [Text]

A list of cost anomaly monitors.

subscribers :: [Subscriber]

A list of subscribers to notify.

threshold :: Double

The dollar value that triggers a notification if the threshold is exceeded.

frequency :: AnomalySubscriptionFrequency

The frequency that anomaly reports are sent over email.

subscriptionName :: Text

The name for the subscription.

newAnomalySubscription

Arguments

:: Double
$sel:threshold:AnomalySubscription'

-> AnomalySubscriptionFrequency
$sel:frequency:AnomalySubscription'

-> Text
$sel:subscriptionName:AnomalySubscription'

-> AnomalySubscription

Create a value of AnomalySubscription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountId:AnomalySubscription', anomalySubscription_accountId - Your unique account identifier.

$sel:subscriptionArn:AnomalySubscription', anomalySubscription_subscriptionArn - The AnomalySubscription Amazon Resource Name (ARN).

$sel:monitorArnList:AnomalySubscription', anomalySubscription_monitorArnList - A list of cost anomaly monitors.

$sel:subscribers:AnomalySubscription', anomalySubscription_subscribers - A list of subscribers to notify.

$sel:threshold:AnomalySubscription', anomalySubscription_threshold - The dollar value that triggers a notification if the threshold is exceeded.

$sel:frequency:AnomalySubscription', anomalySubscription_frequency - The frequency that anomaly reports are sent over email.

$sel:subscriptionName:AnomalySubscription', anomalySubscription_subscriptionName - The name for the subscription.

anomalySubscription_accountId :: Lens' AnomalySubscription (Maybe Text)

Your unique account identifier.

anomalySubscription_subscriptionArn :: Lens' AnomalySubscription (Maybe Text)

The AnomalySubscription Amazon Resource Name (ARN).

anomalySubscription_monitorArnList :: Lens' AnomalySubscription [Text]

A list of cost anomaly monitors.

anomalySubscription_subscribers :: Lens' AnomalySubscription [Subscriber]

A list of subscribers to notify.

anomalySubscription_threshold :: Lens' AnomalySubscription Double

The dollar value that triggers a notification if the threshold is exceeded.

anomalySubscription_frequency :: Lens' AnomalySubscription AnomalySubscriptionFrequency

The frequency that anomaly reports are sent over email.

anomalySubscription_subscriptionName :: Lens' AnomalySubscription Text

The name for the subscription.