Network.AWS.CostExplorer.Types.CostCategory

Description

Documentation

data CostCategory

The structure of Cost Categories. This includes detailed metadata and the set of rules for the CostCategory object.

See: newCostCategory smart constructor.

Constructors

CostCategory'

Fields

splitChargeRules :: Maybe (NonEmpty CostCategorySplitChargeRule)

The split charge rules that are used to allocate your charges between your Cost Category values.

processingStatus :: Maybe [CostCategoryProcessingStatus]

The list of processing statuses for Cost Management products for a specific cost category.

defaultValue :: Maybe Text

effectiveEnd :: Maybe Text

The effective end data of your Cost Category.

costCategoryArn :: Text

The unique identifier for your Cost Category.

effectiveStart :: Text

The effective state data of your Cost Category.

name :: Text

ruleVersion :: CostCategoryRuleVersion

rules :: NonEmpty CostCategoryRule

The rules are processed in order. If there are multiple rules that match the line item, then the first rule to match is used to determine that Cost Category value.

newCostCategory

Arguments

:: Text
$sel:costCategoryArn:CostCategory'

-> Text
$sel:effectiveStart:CostCategory'

-> Text
$sel:name:CostCategory'

-> CostCategoryRuleVersion
$sel:ruleVersion:CostCategory'

-> NonEmpty CostCategoryRule
$sel:rules:CostCategory'

-> CostCategory

Create a value of CostCategory with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:splitChargeRules:CostCategory', costCategory_splitChargeRules - The split charge rules that are used to allocate your charges between your Cost Category values.

$sel:processingStatus:CostCategory', costCategory_processingStatus - The list of processing statuses for Cost Management products for a specific cost category.

$sel:defaultValue:CostCategory', costCategory_defaultValue - Undocumented member.

$sel:effectiveEnd:CostCategory', costCategory_effectiveEnd - The effective end data of your Cost Category.

$sel:costCategoryArn:CostCategory', costCategory_costCategoryArn - The unique identifier for your Cost Category.

$sel:effectiveStart:CostCategory', costCategory_effectiveStart - The effective state data of your Cost Category.

$sel:name:CostCategory', costCategory_name - Undocumented member.

$sel:ruleVersion:CostCategory', costCategory_ruleVersion - Undocumented member.

$sel:rules:CostCategory', costCategory_rules - The rules are processed in order. If there are multiple rules that match the line item, then the first rule to match is used to determine that Cost Category value.

costCategory_splitChargeRules :: Lens' CostCategory (Maybe (NonEmpty CostCategorySplitChargeRule))

The split charge rules that are used to allocate your charges between your Cost Category values.

costCategory_processingStatus :: Lens' CostCategory (Maybe [CostCategoryProcessingStatus])

The list of processing statuses for Cost Management products for a specific cost category.

costCategory_defaultValue :: Lens' CostCategory (Maybe Text)

Undocumented member.

costCategory_effectiveEnd :: Lens' CostCategory (Maybe Text)

The effective end data of your Cost Category.

costCategory_costCategoryArn :: Lens' CostCategory Text

The unique identifier for your Cost Category.

costCategory_effectiveStart :: Lens' CostCategory Text

The effective state data of your Cost Category.

costCategory_name :: Lens' CostCategory Text

Undocumented member.

costCategory_ruleVersion :: Lens' CostCategory CostCategoryRuleVersion

Undocumented member.

costCategory_rules :: Lens' CostCategory (NonEmpty CostCategoryRule)

The rules are processed in order. If there are multiple rules that match the line item, then the first rule to match is used to determine that Cost Category value.