Network.AWS.CostExplorer.Types.CostCategoryValues

Description

Documentation

data CostCategoryValues

The Cost Categories values used for filtering the costs.

If Values and Key are not specified, the ABSENT MatchOption is applied to all Cost Categories. That is, it filters on resources that aren't mapped to any Cost Categories.

If Values is provided and Key isn't specified, the ABSENT MatchOption is applied to the Cost Categories Key only. That is, it filters on resources without the given Cost Categories key.

See: newCostCategoryValues smart constructor.

Constructors

CostCategoryValues'

Fields

values :: Maybe [Text]

The specific value of the Cost Category.

key :: Maybe Text

matchOptions :: Maybe [MatchOption]

The match options that you can use to filter your results. MatchOptions is only applicable for actions related to cost category. The default values for MatchOptions is EQUALS and CASE_SENSITIVE.

newCostCategoryValues :: CostCategoryValues

Create a value of CostCategoryValues with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:values:CostCategoryValues', costCategoryValues_values - The specific value of the Cost Category.

$sel:key:CostCategoryValues', costCategoryValues_key - Undocumented member.

$sel:matchOptions:CostCategoryValues', costCategoryValues_matchOptions - The match options that you can use to filter your results. MatchOptions is only applicable for actions related to cost category. The default values for MatchOptions is EQUALS and CASE_SENSITIVE.

costCategoryValues_values :: Lens' CostCategoryValues (Maybe [Text])

The specific value of the Cost Category.

costCategoryValues_key :: Lens' CostCategoryValues (Maybe Text)

Undocumented member.

costCategoryValues_matchOptions :: Lens' CostCategoryValues (Maybe [MatchOption])

The match options that you can use to filter your results. MatchOptions is only applicable for actions related to cost category. The default values for MatchOptions is EQUALS and CASE_SENSITIVE.