Network.AWS.CostExplorer.Types.Coverage

Description

Documentation

data Coverage

The amount of instance usage that a reservation covered.

See: newCoverage smart constructor.

Constructors

Coverage'

Fields

coverageCost :: Maybe CoverageCost

The amount of cost that the reservation covered.

coverageHours :: Maybe CoverageHours

The amount of instance usage that the reservation covered, in hours.

coverageNormalizedUnits :: Maybe CoverageNormalizedUnits

The amount of instance usage that the reservation covered, in normalized units.

newCoverage :: Coverage

Create a value of Coverage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:coverageCost:Coverage', coverage_coverageCost - The amount of cost that the reservation covered.

$sel:coverageHours:Coverage', coverage_coverageHours - The amount of instance usage that the reservation covered, in hours.

$sel:coverageNormalizedUnits:Coverage', coverage_coverageNormalizedUnits - The amount of instance usage that the reservation covered, in normalized units.

coverage_coverageCost :: Lens' Coverage (Maybe CoverageCost)

The amount of cost that the reservation covered.

coverage_coverageHours :: Lens' Coverage (Maybe CoverageHours)

The amount of instance usage that the reservation covered, in hours.

coverage_coverageNormalizedUnits :: Lens' Coverage (Maybe CoverageNormalizedUnits)

The amount of instance usage that the reservation covered, in normalized units.