Network.AWS.CostExplorer.Types.CoverageNormalizedUnits

Description

Documentation

data CoverageNormalizedUnits

The amount of instance usage, in normalized units. You can use normalized units to see your EC2 usage for multiple sizes of instances in a uniform way. For example, suppose that you run an xlarge instance and a 2xlarge instance. If you run both instances for the same amount of time, the 2xlarge instance uses twice as much of your reservation as the xlarge instance, even though both instances show only one instance-hour. When you use normalized units instead of instance-hours, the xlarge instance used 8 normalized units, and the 2xlarge instance used 16 normalized units.

For more information, see Modifying Reserved Instances in the Amazon Elastic Compute Cloud User Guide for Linux Instances.

See: newCoverageNormalizedUnits smart constructor.

Constructors

CoverageNormalizedUnits'

Fields

reservedNormalizedUnits :: Maybe Text

The number of normalized units that a reservation covers.

totalRunningNormalizedUnits :: Maybe Text

The total number of normalized units that you used.

coverageNormalizedUnitsPercentage :: Maybe Text

The percentage of your used instance normalized units that a reservation covers.

onDemandNormalizedUnits :: Maybe Text

The number of normalized units that are covered by On-Demand Instances instead of a reservation.

newCoverageNormalizedUnits :: CoverageNormalizedUnits

Create a value of CoverageNormalizedUnits with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reservedNormalizedUnits:CoverageNormalizedUnits', coverageNormalizedUnits_reservedNormalizedUnits - The number of normalized units that a reservation covers.

$sel:totalRunningNormalizedUnits:CoverageNormalizedUnits', coverageNormalizedUnits_totalRunningNormalizedUnits - The total number of normalized units that you used.

$sel:coverageNormalizedUnitsPercentage:CoverageNormalizedUnits', coverageNormalizedUnits_coverageNormalizedUnitsPercentage - The percentage of your used instance normalized units that a reservation covers.

$sel:onDemandNormalizedUnits:CoverageNormalizedUnits', coverageNormalizedUnits_onDemandNormalizedUnits - The number of normalized units that are covered by On-Demand Instances instead of a reservation.

coverageNormalizedUnits_reservedNormalizedUnits :: Lens' CoverageNormalizedUnits (Maybe Text)

The number of normalized units that a reservation covers.

coverageNormalizedUnits_totalRunningNormalizedUnits :: Lens' CoverageNormalizedUnits (Maybe Text)

The total number of normalized units that you used.

coverageNormalizedUnits_coverageNormalizedUnitsPercentage :: Lens' CoverageNormalizedUnits (Maybe Text)

The percentage of your used instance normalized units that a reservation covers.

coverageNormalizedUnits_onDemandNormalizedUnits :: Lens' CoverageNormalizedUnits (Maybe Text)

The number of normalized units that are covered by On-Demand Instances instead of a reservation.