Network.AWS.CostExplorer.Types.CurrentInstance

Description

Documentation

data CurrentInstance

Context about the current instance.

See: newCurrentInstance smart constructor.

Constructors

CurrentInstance'

Fields

resourceId :: Maybe Text

Resource ID of the current instance.

savingsPlansCoveredHoursInLookbackPeriod :: Maybe Text

The number of hours during the lookback period that's covered by Savings Plans.

instanceName :: Maybe Text

The name that you given an instance. This field shows as blank if you haven't given the instance a name.

onDemandHoursInLookbackPeriod :: Maybe Text

The number of hours during the lookback period that's billed at On-Demand rates.

currencyCode :: Maybe Text

The currency code that Amazon Web Services used to calculate the costs for this instance.

tags :: Maybe [TagValues]

Cost allocation resource tags that are applied to the instance.

reservationCoveredHoursInLookbackPeriod :: Maybe Text

The number of hours during the lookback period that's covered by reservations.

monthlyCost :: Maybe Text

The current On-Demand cost of operating this instance on a monthly basis.

resourceUtilization :: Maybe ResourceUtilization

Utilization information of the current instance during the lookback period.

resourceDetails :: Maybe ResourceDetails

Details about the resource and utilization.

totalRunningHoursInLookbackPeriod :: Maybe Text

The total number of hours that the instance ran during the lookback period.

newCurrentInstance :: CurrentInstance

Create a value of CurrentInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceId:CurrentInstance', currentInstance_resourceId - Resource ID of the current instance.

$sel:savingsPlansCoveredHoursInLookbackPeriod:CurrentInstance', currentInstance_savingsPlansCoveredHoursInLookbackPeriod - The number of hours during the lookback period that's covered by Savings Plans.

$sel:instanceName:CurrentInstance', currentInstance_instanceName - The name that you given an instance. This field shows as blank if you haven't given the instance a name.

$sel:onDemandHoursInLookbackPeriod:CurrentInstance', currentInstance_onDemandHoursInLookbackPeriod - The number of hours during the lookback period that's billed at On-Demand rates.

$sel:currencyCode:CurrentInstance', currentInstance_currencyCode - The currency code that Amazon Web Services used to calculate the costs for this instance.

$sel:tags:CurrentInstance', currentInstance_tags - Cost allocation resource tags that are applied to the instance.

$sel:reservationCoveredHoursInLookbackPeriod:CurrentInstance', currentInstance_reservationCoveredHoursInLookbackPeriod - The number of hours during the lookback period that's covered by reservations.

$sel:monthlyCost:CurrentInstance', currentInstance_monthlyCost - The current On-Demand cost of operating this instance on a monthly basis.

$sel:resourceUtilization:CurrentInstance', currentInstance_resourceUtilization - Utilization information of the current instance during the lookback period.

$sel:resourceDetails:CurrentInstance', currentInstance_resourceDetails - Details about the resource and utilization.

$sel:totalRunningHoursInLookbackPeriod:CurrentInstance', currentInstance_totalRunningHoursInLookbackPeriod - The total number of hours that the instance ran during the lookback period.

currentInstance_resourceId :: Lens' CurrentInstance (Maybe Text)

Resource ID of the current instance.

currentInstance_savingsPlansCoveredHoursInLookbackPeriod :: Lens' CurrentInstance (Maybe Text)

The number of hours during the lookback period that's covered by Savings Plans.

currentInstance_instanceName :: Lens' CurrentInstance (Maybe Text)

The name that you given an instance. This field shows as blank if you haven't given the instance a name.

currentInstance_onDemandHoursInLookbackPeriod :: Lens' CurrentInstance (Maybe Text)

The number of hours during the lookback period that's billed at On-Demand rates.

currentInstance_currencyCode :: Lens' CurrentInstance (Maybe Text)

The currency code that Amazon Web Services used to calculate the costs for this instance.

currentInstance_tags :: Lens' CurrentInstance (Maybe [TagValues])

Cost allocation resource tags that are applied to the instance.

currentInstance_reservationCoveredHoursInLookbackPeriod :: Lens' CurrentInstance (Maybe Text)

The number of hours during the lookback period that's covered by reservations.

currentInstance_monthlyCost :: Lens' CurrentInstance (Maybe Text)

The current On-Demand cost of operating this instance on a monthly basis.

currentInstance_resourceUtilization :: Lens' CurrentInstance (Maybe ResourceUtilization)

Utilization information of the current instance during the lookback period.

currentInstance_resourceDetails :: Lens' CurrentInstance (Maybe ResourceDetails)

Details about the resource and utilization.

currentInstance_totalRunningHoursInLookbackPeriod :: Lens' CurrentInstance (Maybe Text)

The total number of hours that the instance ran during the lookback period.