Network.AWS.CostExplorer.Types.EC2InstanceDetails

Description

Documentation

data EC2InstanceDetails

Details about the Amazon EC2 instances that Amazon Web Services recommends that you purchase.

See: newEC2InstanceDetails smart constructor.

Constructors

EC2InstanceDetails'

Fields

platform :: Maybe Text

The platform of the recommended reservation. The platform is the specific combination of operating system, license model, and software on an instance.

instanceType :: Maybe Text

The type of instance that Amazon Web Services recommends.

tenancy :: Maybe Text

Determines whether the recommended reservation is dedicated or shared.

sizeFlexEligible :: Maybe Bool

Determines whether the recommended reservation is size flexible.

currentGeneration :: Maybe Bool

Determines whether the recommendation is for a current-generation instance.

availabilityZone :: Maybe Text

The Availability Zone of the recommended reservation.

family :: Maybe Text

The instance family of the recommended reservation.

region :: Maybe Text

The Amazon Web Services Region of the recommended reservation.

newEC2InstanceDetails :: EC2InstanceDetails

Create a value of EC2InstanceDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:platform:EC2InstanceDetails', eC2InstanceDetails_platform - The platform of the recommended reservation. The platform is the specific combination of operating system, license model, and software on an instance.

$sel:instanceType:EC2InstanceDetails', eC2InstanceDetails_instanceType - The type of instance that Amazon Web Services recommends.

$sel:tenancy:EC2InstanceDetails', eC2InstanceDetails_tenancy - Determines whether the recommended reservation is dedicated or shared.

$sel:sizeFlexEligible:EC2InstanceDetails', eC2InstanceDetails_sizeFlexEligible - Determines whether the recommended reservation is size flexible.

$sel:currentGeneration:EC2InstanceDetails', eC2InstanceDetails_currentGeneration - Determines whether the recommendation is for a current-generation instance.

$sel:availabilityZone:EC2InstanceDetails', eC2InstanceDetails_availabilityZone - The Availability Zone of the recommended reservation.

$sel:family:EC2InstanceDetails', eC2InstanceDetails_family - The instance family of the recommended reservation.

$sel:region:EC2InstanceDetails', eC2InstanceDetails_region - The Amazon Web Services Region of the recommended reservation.

eC2InstanceDetails_platform :: Lens' EC2InstanceDetails (Maybe Text)

The platform of the recommended reservation. The platform is the specific combination of operating system, license model, and software on an instance.

eC2InstanceDetails_instanceType :: Lens' EC2InstanceDetails (Maybe Text)

The type of instance that Amazon Web Services recommends.

eC2InstanceDetails_tenancy :: Lens' EC2InstanceDetails (Maybe Text)

Determines whether the recommended reservation is dedicated or shared.

eC2InstanceDetails_sizeFlexEligible :: Lens' EC2InstanceDetails (Maybe Bool)

Determines whether the recommended reservation is size flexible.

eC2InstanceDetails_currentGeneration :: Lens' EC2InstanceDetails (Maybe Bool)

Determines whether the recommendation is for a current-generation instance.

eC2InstanceDetails_availabilityZone :: Lens' EC2InstanceDetails (Maybe Text)

The Availability Zone of the recommended reservation.

eC2InstanceDetails_family :: Lens' EC2InstanceDetails (Maybe Text)

The instance family of the recommended reservation.

eC2InstanceDetails_region :: Lens' EC2InstanceDetails (Maybe Text)

The Amazon Web Services Region of the recommended reservation.