Network.AWS.CostExplorer.Types.EC2ResourceDetails

Description

Documentation

data EC2ResourceDetails

Details on the Amazon EC2 Resource.

See: newEC2ResourceDetails smart constructor.

Constructors

EC2ResourceDetails'

Fields

platform :: Maybe Text

The platform of the Amazon Web Services instance. The platform is the specific combination of operating system, license model, and software on an instance.

memory :: Maybe Text

The memory capacity of the Amazon Web Services instance.

instanceType :: Maybe Text

The type of Amazon Web Services instance.

vcpu :: Maybe Text

The number of VCPU cores in the Amazon Web Services instance type.

storage :: Maybe Text

The disk storage of the Amazon Web Services instance. This doesn't include EBS storage.

hourlyOnDemandRate :: Maybe Text

The hourly public On-Demand rate for the instance type.

networkPerformance :: Maybe Text

The network performance capacity of the Amazon Web Services instance.

region :: Maybe Text

The Amazon Web Services Region of the instance.

sku :: Maybe Text

The SKU of the product.

newEC2ResourceDetails :: EC2ResourceDetails

Create a value of EC2ResourceDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:platform:EC2ResourceDetails', eC2ResourceDetails_platform - The platform of the Amazon Web Services instance. The platform is the specific combination of operating system, license model, and software on an instance.

$sel:memory:EC2ResourceDetails', eC2ResourceDetails_memory - The memory capacity of the Amazon Web Services instance.

$sel:instanceType:EC2ResourceDetails', eC2ResourceDetails_instanceType - The type of Amazon Web Services instance.

$sel:vcpu:EC2ResourceDetails', eC2ResourceDetails_vcpu - The number of VCPU cores in the Amazon Web Services instance type.

$sel:storage:EC2ResourceDetails', eC2ResourceDetails_storage - The disk storage of the Amazon Web Services instance. This doesn't include EBS storage.

$sel:hourlyOnDemandRate:EC2ResourceDetails', eC2ResourceDetails_hourlyOnDemandRate - The hourly public On-Demand rate for the instance type.

$sel:networkPerformance:EC2ResourceDetails', eC2ResourceDetails_networkPerformance - The network performance capacity of the Amazon Web Services instance.

$sel:region:EC2ResourceDetails', eC2ResourceDetails_region - The Amazon Web Services Region of the instance.

$sel:sku:EC2ResourceDetails', eC2ResourceDetails_sku - The SKU of the product.

eC2ResourceDetails_platform :: Lens' EC2ResourceDetails (Maybe Text)

The platform of the Amazon Web Services instance. The platform is the specific combination of operating system, license model, and software on an instance.

eC2ResourceDetails_memory :: Lens' EC2ResourceDetails (Maybe Text)

The memory capacity of the Amazon Web Services instance.

eC2ResourceDetails_instanceType :: Lens' EC2ResourceDetails (Maybe Text)

The type of Amazon Web Services instance.

eC2ResourceDetails_vcpu :: Lens' EC2ResourceDetails (Maybe Text)

The number of VCPU cores in the Amazon Web Services instance type.

eC2ResourceDetails_storage :: Lens' EC2ResourceDetails (Maybe Text)

The disk storage of the Amazon Web Services instance. This doesn't include EBS storage.

eC2ResourceDetails_hourlyOnDemandRate :: Lens' EC2ResourceDetails (Maybe Text)

The hourly public On-Demand rate for the instance type.

eC2ResourceDetails_networkPerformance :: Lens' EC2ResourceDetails (Maybe Text)

The network performance capacity of the Amazon Web Services instance.

eC2ResourceDetails_region :: Lens' EC2ResourceDetails (Maybe Text)

The Amazon Web Services Region of the instance.

eC2ResourceDetails_sku :: Lens' EC2ResourceDetails (Maybe Text)

The SKU of the product.