Network.AWS.CostExplorer.Types.ESInstanceDetails

Description

Documentation

data ESInstanceDetails

Details about the Amazon ES instances that Amazon Web Services recommends that you purchase.

See: newESInstanceDetails smart constructor.

Constructors

ESInstanceDetails'

Fields

sizeFlexEligible :: Maybe Bool

Determines whether the recommended reservation is size flexible.

currentGeneration :: Maybe Bool

Determines whether the recommendation is for a current-generation instance.

instanceClass :: Maybe Text

The class of instance that Amazon Web Services recommends.

instanceSize :: Maybe Text

The size of instance that Amazon Web Services recommends.

region :: Maybe Text

The Amazon Web Services Region of the recommended reservation.

newESInstanceDetails :: ESInstanceDetails

Create a value of ESInstanceDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sizeFlexEligible:ESInstanceDetails', eSInstanceDetails_sizeFlexEligible - Determines whether the recommended reservation is size flexible.

$sel:currentGeneration:ESInstanceDetails', eSInstanceDetails_currentGeneration - Determines whether the recommendation is for a current-generation instance.

$sel:instanceClass:ESInstanceDetails', eSInstanceDetails_instanceClass - The class of instance that Amazon Web Services recommends.

$sel:instanceSize:ESInstanceDetails', eSInstanceDetails_instanceSize - The size of instance that Amazon Web Services recommends.

$sel:region:ESInstanceDetails', eSInstanceDetails_region - The Amazon Web Services Region of the recommended reservation.

eSInstanceDetails_sizeFlexEligible :: Lens' ESInstanceDetails (Maybe Bool)

Determines whether the recommended reservation is size flexible.

eSInstanceDetails_currentGeneration :: Lens' ESInstanceDetails (Maybe Bool)

Determines whether the recommendation is for a current-generation instance.

eSInstanceDetails_instanceClass :: Lens' ESInstanceDetails (Maybe Text)

The class of instance that Amazon Web Services recommends.

eSInstanceDetails_instanceSize :: Lens' ESInstanceDetails (Maybe Text)

The size of instance that Amazon Web Services recommends.

eSInstanceDetails_region :: Lens' ESInstanceDetails (Maybe Text)

The Amazon Web Services Region of the recommended reservation.