Network.AWS.CostExplorer.Types.Expression

Description

Documentation

data Expression

Use Expression to filter by cost or by usage. There are two patterns:

Simple dimension values - You can set the dimension name and values for the filters that you plan to use. For example, you can filter for REGION==us-east-1 OR REGION==us-west-1. For GetRightsizingRecommendation, the Region is a full name (for example, REGION==US East (N. Virginia). The Expression example is as follows:

{ "Dimensions": { "Key": "REGION", "Values": [ "us-east-1", “us-west-1” ] } }

The list of dimension values are OR'd together to retrieve cost or usage data. You can create Expression and DimensionValues objects using either with* methods or set* methods in multiple lines.

Compound dimension values with logical operations - You can use multiple Expression types and the logical operators AND/OR/NOT to create a list of one or more Expression objects. By doing this, you can filter on more advanced options. For example, you can filter on ((REGION == us-east-1 OR REGION == us-west-1) OR (TAG.Type == Type1)) AND (USAGE_TYPE != DataTransfer). The Expression for that is as follows:

{ "And": [ {"Or": [ {"Dimensions": { "Key": "REGION", "Values": [ "us-east-1", "us-west-1" ] }}, {"Tags": { "Key": "TagName", "Values": ["Value1"] } } ]}, {"Not": {"Dimensions": { "Key": "USAGE_TYPE", "Values": ["DataTransfer"] }}} ] }

Because each Expression can have only one operator, the service returns an error if more than one is specified. The following example shows an Expression object that creates an error.

 { "And": [ ... ], "DimensionValues": { "Dimension": "USAGE_TYPE", "Values": [ "DataTransfer" ] } }

For the GetRightsizingRecommendation action, a combination of OR and NOT isn't supported. OR isn't supported between different dimensions, or dimensions and tags. NOT operators aren't supported. Dimensions are also limited to LINKED_ACCOUNT, REGION, or RIGHTSIZING_TYPE.

For the GetReservationPurchaseRecommendation action, only NOT is supported. AND and OR aren't supported. Dimensions are limited to LINKED_ACCOUNT.

See: newExpression smart constructor.

Constructors

Expression'

Fields

costCategories :: Maybe CostCategoryValues

The filter that's based on CostCategory values.

not :: Maybe Expression

Return results that don't match a Dimension object.

or :: Maybe [Expression]

Return results that match either Dimension object.

tags :: Maybe TagValues

The specific Tag to use for Expression.

and :: Maybe [Expression]

Return results that match both Dimension objects.

dimensions :: Maybe DimensionValues

The specific Dimension to use for Expression.

newExpression :: Expression

Create a value of Expression with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:costCategories:Expression', expression_costCategories - The filter that's based on CostCategory values.

$sel:not:Expression', expression_not - Return results that don't match a Dimension object.

$sel:or:Expression', expression_or - Return results that match either Dimension object.

$sel:tags:Expression', expression_tags - The specific Tag to use for Expression.

$sel:and:Expression', expression_and - Return results that match both Dimension objects.

$sel:dimensions:Expression', expression_dimensions - The specific Dimension to use for Expression.

expression_costCategories :: Lens' Expression (Maybe CostCategoryValues)

The filter that's based on CostCategory values.

expression_not :: Lens' Expression (Maybe Expression)

Return results that don't match a Dimension object.

expression_or :: Lens' Expression (Maybe [Expression])

Return results that match either Dimension object.

expression_tags :: Lens' Expression (Maybe TagValues)

The specific Tag to use for Expression.

expression_and :: Lens' Expression (Maybe [Expression])

Return results that match both Dimension objects.

expression_dimensions :: Lens' Expression (Maybe DimensionValues)

The specific Dimension to use for Expression.