Network.AWS.CostExplorer.Types.ReservationPurchaseRecommendation

Description

Documentation

data ReservationPurchaseRecommendation

A specific reservation that Amazon Web Services recommends for purchase.

See: newReservationPurchaseRecommendation smart constructor.

Constructors

ReservationPurchaseRecommendation'

Fields

termInYears :: Maybe TermInYears

The term of the reservation that you want recommendations for, in years.

recommendationSummary :: Maybe ReservationPurchaseRecommendationSummary

A summary about the recommended purchase.

serviceSpecification :: Maybe ServiceSpecification

Hardware specifications for the service that you want recommendations for.

accountScope :: Maybe AccountScope

The account scope that Amazon Web Services recommends that you purchase this instance for. For example, you can purchase this reservation for an entire organization in Amazon Web Services Organizations.

recommendationDetails :: Maybe [ReservationPurchaseRecommendationDetail]

Details about the recommended purchases.

lookbackPeriodInDays :: Maybe LookbackPeriodInDays

How many days of previous usage that Amazon Web Services considers when making this recommendation.

paymentOption :: Maybe PaymentOption

The payment option for the reservation (for example, AllUpfront or NoUpfront).

newReservationPurchaseRecommendation :: ReservationPurchaseRecommendation

Create a value of ReservationPurchaseRecommendation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:termInYears:ReservationPurchaseRecommendation', reservationPurchaseRecommendation_termInYears - The term of the reservation that you want recommendations for, in years.

$sel:recommendationSummary:ReservationPurchaseRecommendation', reservationPurchaseRecommendation_recommendationSummary - A summary about the recommended purchase.

$sel:serviceSpecification:ReservationPurchaseRecommendation', reservationPurchaseRecommendation_serviceSpecification - Hardware specifications for the service that you want recommendations for.

$sel:accountScope:ReservationPurchaseRecommendation', reservationPurchaseRecommendation_accountScope - The account scope that Amazon Web Services recommends that you purchase this instance for. For example, you can purchase this reservation for an entire organization in Amazon Web Services Organizations.

$sel:recommendationDetails:ReservationPurchaseRecommendation', reservationPurchaseRecommendation_recommendationDetails - Details about the recommended purchases.

$sel:lookbackPeriodInDays:ReservationPurchaseRecommendation', reservationPurchaseRecommendation_lookbackPeriodInDays - How many days of previous usage that Amazon Web Services considers when making this recommendation.

$sel:paymentOption:ReservationPurchaseRecommendation', reservationPurchaseRecommendation_paymentOption - The payment option for the reservation (for example, AllUpfront or NoUpfront).

reservationPurchaseRecommendation_termInYears :: Lens' ReservationPurchaseRecommendation (Maybe TermInYears)

The term of the reservation that you want recommendations for, in years.

reservationPurchaseRecommendation_recommendationSummary :: Lens' ReservationPurchaseRecommendation (Maybe ReservationPurchaseRecommendationSummary)

A summary about the recommended purchase.

reservationPurchaseRecommendation_serviceSpecification :: Lens' ReservationPurchaseRecommendation (Maybe ServiceSpecification)

Hardware specifications for the service that you want recommendations for.

reservationPurchaseRecommendation_accountScope :: Lens' ReservationPurchaseRecommendation (Maybe AccountScope)

The account scope that Amazon Web Services recommends that you purchase this instance for. For example, you can purchase this reservation for an entire organization in Amazon Web Services Organizations.

reservationPurchaseRecommendation_recommendationDetails :: Lens' ReservationPurchaseRecommendation (Maybe [ReservationPurchaseRecommendationDetail])

Details about the recommended purchases.

reservationPurchaseRecommendation_lookbackPeriodInDays :: Lens' ReservationPurchaseRecommendation (Maybe LookbackPeriodInDays)

How many days of previous usage that Amazon Web Services considers when making this recommendation.

reservationPurchaseRecommendation_paymentOption :: Lens' ReservationPurchaseRecommendation (Maybe PaymentOption)

The payment option for the reservation (for example, AllUpfront or NoUpfront).