Network.AWS.CostExplorer.Types.ResultByTime

Description

Documentation

data ResultByTime

The result that's associated with a time period.

See: newResultByTime smart constructor.

Constructors

ResultByTime'

Fields

groups :: Maybe [Group]

The groups that this time period includes.

timePeriod :: Maybe DateInterval

The time period that the result covers.

total :: Maybe (HashMap Text MetricValue)

The total amount of cost or usage accrued during the time period.

estimated :: Maybe Bool

Determines whether the result is estimated.

newResultByTime :: ResultByTime

Create a value of ResultByTime with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groups:ResultByTime', resultByTime_groups - The groups that this time period includes.

$sel:timePeriod:ResultByTime', resultByTime_timePeriod - The time period that the result covers.

$sel:total:ResultByTime', resultByTime_total - The total amount of cost or usage accrued during the time period.

$sel:estimated:ResultByTime', resultByTime_estimated - Determines whether the result is estimated.

resultByTime_groups :: Lens' ResultByTime (Maybe [Group])

The groups that this time period includes.

resultByTime_timePeriod :: Lens' ResultByTime (Maybe DateInterval)

The time period that the result covers.

resultByTime_total :: Lens' ResultByTime (Maybe (HashMap Text MetricValue))

The total amount of cost or usage accrued during the time period.

resultByTime_estimated :: Lens' ResultByTime (Maybe Bool)

Determines whether the result is estimated.