Network.AWS.CostExplorer.Types.RightsizingRecommendation

Description

Documentation

data RightsizingRecommendation

Recommendations to rightsize resources.

See: newRightsizingRecommendation smart constructor.

Constructors

RightsizingRecommendation'

Fields

accountId :: Maybe Text

The account that this recommendation is for.

terminateRecommendationDetail :: Maybe TerminateRecommendationDetail

The details for termination recommendations.

currentInstance :: Maybe CurrentInstance

Context regarding the current instance.

rightsizingType :: Maybe RightsizingType

A recommendation to either terminate or modify the resource.

findingReasonCodes :: Maybe [FindingReasonCode]

The list of possible reasons why the recommendation is generated such as under or over utilization of specific metrics (for example, CPU, Memory, Network).

modifyRecommendationDetail :: Maybe ModifyRecommendationDetail

The details for the modification recommendations.

newRightsizingRecommendation :: RightsizingRecommendation

Create a value of RightsizingRecommendation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountId:RightsizingRecommendation', rightsizingRecommendation_accountId - The account that this recommendation is for.

$sel:terminateRecommendationDetail:RightsizingRecommendation', rightsizingRecommendation_terminateRecommendationDetail - The details for termination recommendations.

$sel:currentInstance:RightsizingRecommendation', rightsizingRecommendation_currentInstance - Context regarding the current instance.

$sel:rightsizingType:RightsizingRecommendation', rightsizingRecommendation_rightsizingType - A recommendation to either terminate or modify the resource.

$sel:findingReasonCodes:RightsizingRecommendation', rightsizingRecommendation_findingReasonCodes - The list of possible reasons why the recommendation is generated such as under or over utilization of specific metrics (for example, CPU, Memory, Network).

$sel:modifyRecommendationDetail:RightsizingRecommendation', rightsizingRecommendation_modifyRecommendationDetail - The details for the modification recommendations.

rightsizingRecommendation_accountId :: Lens' RightsizingRecommendation (Maybe Text)

The account that this recommendation is for.

rightsizingRecommendation_terminateRecommendationDetail :: Lens' RightsizingRecommendation (Maybe TerminateRecommendationDetail)

The details for termination recommendations.

rightsizingRecommendation_currentInstance :: Lens' RightsizingRecommendation (Maybe CurrentInstance)

Context regarding the current instance.

rightsizingRecommendation_rightsizingType :: Lens' RightsizingRecommendation (Maybe RightsizingType)

A recommendation to either terminate or modify the resource.

rightsizingRecommendation_findingReasonCodes :: Lens' RightsizingRecommendation (Maybe [FindingReasonCode])

The list of possible reasons why the recommendation is generated such as under or over utilization of specific metrics (for example, CPU, Memory, Network).

rightsizingRecommendation_modifyRecommendationDetail :: Lens' RightsizingRecommendation (Maybe ModifyRecommendationDetail)

The details for the modification recommendations.