Network.AWS.CostExplorer.Types.RightsizingRecommendationSummary

Description

Documentation

data RightsizingRecommendationSummary

The summary of rightsizing recommendations

See: newRightsizingRecommendationSummary smart constructor.

Constructors

RightsizingRecommendationSummary'

Fields

savingsPercentage :: Maybe Text

The savings percentage based on the recommended modifications. It's relative to the total On-Demand costs that are associated with these instances.

savingsCurrencyCode :: Maybe Text

The currency code that Amazon Web Services used to calculate the savings.

totalRecommendationCount :: Maybe Text

The total number of instance recommendations.

estimatedTotalMonthlySavingsAmount :: Maybe Text

The estimated total savings resulting from modifications, on a monthly basis.

newRightsizingRecommendationSummary :: RightsizingRecommendationSummary

Create a value of RightsizingRecommendationSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:savingsPercentage:RightsizingRecommendationSummary', rightsizingRecommendationSummary_savingsPercentage - The savings percentage based on the recommended modifications. It's relative to the total On-Demand costs that are associated with these instances.

$sel:savingsCurrencyCode:RightsizingRecommendationSummary', rightsizingRecommendationSummary_savingsCurrencyCode - The currency code that Amazon Web Services used to calculate the savings.

$sel:totalRecommendationCount:RightsizingRecommendationSummary', rightsizingRecommendationSummary_totalRecommendationCount - The total number of instance recommendations.

$sel:estimatedTotalMonthlySavingsAmount:RightsizingRecommendationSummary', rightsizingRecommendationSummary_estimatedTotalMonthlySavingsAmount - The estimated total savings resulting from modifications, on a monthly basis.

rightsizingRecommendationSummary_savingsPercentage :: Lens' RightsizingRecommendationSummary (Maybe Text)

The savings percentage based on the recommended modifications. It's relative to the total On-Demand costs that are associated with these instances.

rightsizingRecommendationSummary_savingsCurrencyCode :: Lens' RightsizingRecommendationSummary (Maybe Text)

The currency code that Amazon Web Services used to calculate the savings.

rightsizingRecommendationSummary_totalRecommendationCount :: Lens' RightsizingRecommendationSummary (Maybe Text)

The total number of instance recommendations.

rightsizingRecommendationSummary_estimatedTotalMonthlySavingsAmount :: Lens' RightsizingRecommendationSummary (Maybe Text)

The estimated total savings resulting from modifications, on a monthly basis.