Network.AWS.CostExplorer.Types.SavingsPlansCoverage

Description

Documentation

data SavingsPlansCoverage

The amount of Savings Plans eligible usage that is covered by Savings Plans. All calculations consider the On-Demand equivalent of your Savings Plans usage.

See: newSavingsPlansCoverage smart constructor.

Constructors

SavingsPlansCoverage'

Fields

timePeriod :: Maybe DateInterval

attributes :: Maybe (HashMap Text Text)

The attribute that applies to a specific Dimension.

coverage :: Maybe SavingsPlansCoverageData

The amount of Savings Plans eligible usage that the Savings Plans covered.

newSavingsPlansCoverage :: SavingsPlansCoverage

Create a value of SavingsPlansCoverage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timePeriod:SavingsPlansCoverage', savingsPlansCoverage_timePeriod - Undocumented member.

$sel:attributes:SavingsPlansCoverage', savingsPlansCoverage_attributes - The attribute that applies to a specific Dimension.

$sel:coverage:SavingsPlansCoverage', savingsPlansCoverage_coverage - The amount of Savings Plans eligible usage that the Savings Plans covered.

savingsPlansCoverage_timePeriod :: Lens' SavingsPlansCoverage (Maybe DateInterval)

Undocumented member.

savingsPlansCoverage_attributes :: Lens' SavingsPlansCoverage (Maybe (HashMap Text Text))

The attribute that applies to a specific Dimension.

savingsPlansCoverage_coverage :: Lens' SavingsPlansCoverage (Maybe SavingsPlansCoverageData)

The amount of Savings Plans eligible usage that the Savings Plans covered.