Network.AWS.CostExplorer.Types.SavingsPlansPurchaseRecommendation

Description

Documentation

data SavingsPlansPurchaseRecommendation

Contains your request parameters, Savings Plan Recommendations Summary, and Details.

See: newSavingsPlansPurchaseRecommendation smart constructor.

Constructors

SavingsPlansPurchaseRecommendation'

Fields

paymentOption :: Maybe PaymentOption

The payment option used to generate the recommendation.

savingsPlansPurchaseRecommendationSummary :: Maybe SavingsPlansPurchaseRecommendationSummary

Summary metrics for your Savings Plans Recommendations.

accountScope :: Maybe AccountScope

The account scope that you want your recommendations for. Amazon Web Services calculates recommendations that include the management account and member accounts if the value is set to PAYER. If the value is LINKED, recommendations are calculated for individual member accounts only.

termInYears :: Maybe TermInYears

The Savings Plans recommendation term in years. It's used to generate the recommendation.

savingsPlansType :: Maybe SupportedSavingsPlansType

The requested Savings Plans recommendation type.

savingsPlansPurchaseRecommendationDetails :: Maybe [SavingsPlansPurchaseRecommendationDetail]

Details for the Savings Plans we recommend that you purchase to cover existing Savings Plans eligible workloads.

lookbackPeriodInDays :: Maybe LookbackPeriodInDays

The lookback period in days, used to generate the recommendation.

newSavingsPlansPurchaseRecommendation :: SavingsPlansPurchaseRecommendation

Create a value of SavingsPlansPurchaseRecommendation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:paymentOption:SavingsPlansPurchaseRecommendation', savingsPlansPurchaseRecommendation_paymentOption - The payment option used to generate the recommendation.

$sel:savingsPlansPurchaseRecommendationSummary:SavingsPlansPurchaseRecommendation', savingsPlansPurchaseRecommendation_savingsPlansPurchaseRecommendationSummary - Summary metrics for your Savings Plans Recommendations.

$sel:accountScope:SavingsPlansPurchaseRecommendation', savingsPlansPurchaseRecommendation_accountScope - The account scope that you want your recommendations for. Amazon Web Services calculates recommendations that include the management account and member accounts if the value is set to PAYER. If the value is LINKED, recommendations are calculated for individual member accounts only.

$sel:termInYears:SavingsPlansPurchaseRecommendation', savingsPlansPurchaseRecommendation_termInYears - The Savings Plans recommendation term in years. It's used to generate the recommendation.

$sel:savingsPlansType:SavingsPlansPurchaseRecommendation', savingsPlansPurchaseRecommendation_savingsPlansType - The requested Savings Plans recommendation type.

$sel:savingsPlansPurchaseRecommendationDetails:SavingsPlansPurchaseRecommendation', savingsPlansPurchaseRecommendation_savingsPlansPurchaseRecommendationDetails - Details for the Savings Plans we recommend that you purchase to cover existing Savings Plans eligible workloads.

$sel:lookbackPeriodInDays:SavingsPlansPurchaseRecommendation', savingsPlansPurchaseRecommendation_lookbackPeriodInDays - The lookback period in days, used to generate the recommendation.

savingsPlansPurchaseRecommendation_paymentOption :: Lens' SavingsPlansPurchaseRecommendation (Maybe PaymentOption)

The payment option used to generate the recommendation.

savingsPlansPurchaseRecommendation_savingsPlansPurchaseRecommendationSummary :: Lens' SavingsPlansPurchaseRecommendation (Maybe SavingsPlansPurchaseRecommendationSummary)

Summary metrics for your Savings Plans Recommendations.

savingsPlansPurchaseRecommendation_accountScope :: Lens' SavingsPlansPurchaseRecommendation (Maybe AccountScope)

The account scope that you want your recommendations for. Amazon Web Services calculates recommendations that include the management account and member accounts if the value is set to PAYER. If the value is LINKED, recommendations are calculated for individual member accounts only.

savingsPlansPurchaseRecommendation_termInYears :: Lens' SavingsPlansPurchaseRecommendation (Maybe TermInYears)

The Savings Plans recommendation term in years. It's used to generate the recommendation.

savingsPlansPurchaseRecommendation_savingsPlansType :: Lens' SavingsPlansPurchaseRecommendation (Maybe SupportedSavingsPlansType)

The requested Savings Plans recommendation type.

savingsPlansPurchaseRecommendation_savingsPlansPurchaseRecommendationDetails :: Lens' SavingsPlansPurchaseRecommendation (Maybe [SavingsPlansPurchaseRecommendationDetail])

Details for the Savings Plans we recommend that you purchase to cover existing Savings Plans eligible workloads.

savingsPlansPurchaseRecommendation_lookbackPeriodInDays :: Lens' SavingsPlansPurchaseRecommendation (Maybe LookbackPeriodInDays)

The lookback period in days, used to generate the recommendation.