Network.AWS.CostExplorer.Types.SavingsPlansPurchaseRecommendationSummary

Description

Documentation

data SavingsPlansPurchaseRecommendationSummary

Summary metrics for your Savings Plans Purchase Recommendations.

See: newSavingsPlansPurchaseRecommendationSummary smart constructor.

Constructors

SavingsPlansPurchaseRecommendationSummary'

Fields

currencyCode :: Maybe Text

The currency code that Amazon Web Services used to generate the recommendations and present potential savings.

dailyCommitmentToPurchase :: Maybe Text

The recommended Savings Plans cost on a daily (24 hourly) basis.

estimatedTotalCost :: Maybe Text

The estimated total cost of the usage after purchasing the recommended Savings Plans. This is a sum of the cost of Savings Plans during this term, and the remaining On-Demand usage.

estimatedROI :: Maybe Text

The estimated return on investment that's based on the recommended Savings Plans and estimated savings.

estimatedSavingsAmount :: Maybe Text

The estimated total savings over the lookback period, based on the purchase of the recommended Savings Plans.

estimatedMonthlySavingsAmount :: Maybe Text

The estimated monthly savings amount that's based on the recommended Savings Plans purchase.

estimatedOnDemandCostWithCurrentCommitment :: Maybe Text

The estimated On-Demand costs you would expect with no additional commitment. It's based on your usage of the selected time period and the Savings Plans you own.

estimatedSavingsPercentage :: Maybe Text

The estimated savings relative to the total cost of On-Demand usage, over the lookback period. This is calculated as estimatedSavingsAmount/ CurrentOnDemandSpend*100.

totalRecommendationCount :: Maybe Text

The aggregate number of Savings Plans recommendations that exist for your account.

currentOnDemandSpend :: Maybe Text

The current total on demand spend of the applicable usage types over the lookback period.

hourlyCommitmentToPurchase :: Maybe Text

The recommended hourly commitment that's based on the recommendation parameters.

newSavingsPlansPurchaseRecommendationSummary :: SavingsPlansPurchaseRecommendationSummary

Create a value of SavingsPlansPurchaseRecommendationSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:currencyCode:SavingsPlansPurchaseRecommendationSummary', savingsPlansPurchaseRecommendationSummary_currencyCode - The currency code that Amazon Web Services used to generate the recommendations and present potential savings.

$sel:dailyCommitmentToPurchase:SavingsPlansPurchaseRecommendationSummary', savingsPlansPurchaseRecommendationSummary_dailyCommitmentToPurchase - The recommended Savings Plans cost on a daily (24 hourly) basis.

$sel:estimatedTotalCost:SavingsPlansPurchaseRecommendationSummary', savingsPlansPurchaseRecommendationSummary_estimatedTotalCost - The estimated total cost of the usage after purchasing the recommended Savings Plans. This is a sum of the cost of Savings Plans during this term, and the remaining On-Demand usage.

$sel:estimatedROI:SavingsPlansPurchaseRecommendationSummary', savingsPlansPurchaseRecommendationSummary_estimatedROI - The estimated return on investment that's based on the recommended Savings Plans and estimated savings.

$sel:estimatedSavingsAmount:SavingsPlansPurchaseRecommendationSummary', savingsPlansPurchaseRecommendationSummary_estimatedSavingsAmount - The estimated total savings over the lookback period, based on the purchase of the recommended Savings Plans.

$sel:estimatedMonthlySavingsAmount:SavingsPlansPurchaseRecommendationSummary', savingsPlansPurchaseRecommendationSummary_estimatedMonthlySavingsAmount - The estimated monthly savings amount that's based on the recommended Savings Plans purchase.

$sel:estimatedOnDemandCostWithCurrentCommitment:SavingsPlansPurchaseRecommendationSummary', savingsPlansPurchaseRecommendationSummary_estimatedOnDemandCostWithCurrentCommitment - The estimated On-Demand costs you would expect with no additional commitment. It's based on your usage of the selected time period and the Savings Plans you own.

$sel:estimatedSavingsPercentage:SavingsPlansPurchaseRecommendationSummary', savingsPlansPurchaseRecommendationSummary_estimatedSavingsPercentage - The estimated savings relative to the total cost of On-Demand usage, over the lookback period. This is calculated as estimatedSavingsAmount/ CurrentOnDemandSpend*100.

$sel:totalRecommendationCount:SavingsPlansPurchaseRecommendationSummary', savingsPlansPurchaseRecommendationSummary_totalRecommendationCount - The aggregate number of Savings Plans recommendations that exist for your account.

$sel:currentOnDemandSpend:SavingsPlansPurchaseRecommendationSummary', savingsPlansPurchaseRecommendationSummary_currentOnDemandSpend - The current total on demand spend of the applicable usage types over the lookback period.

$sel:hourlyCommitmentToPurchase:SavingsPlansPurchaseRecommendationSummary', savingsPlansPurchaseRecommendationSummary_hourlyCommitmentToPurchase - The recommended hourly commitment that's based on the recommendation parameters.

savingsPlansPurchaseRecommendationSummary_currencyCode :: Lens' SavingsPlansPurchaseRecommendationSummary (Maybe Text)

The currency code that Amazon Web Services used to generate the recommendations and present potential savings.

savingsPlansPurchaseRecommendationSummary_dailyCommitmentToPurchase :: Lens' SavingsPlansPurchaseRecommendationSummary (Maybe Text)

The recommended Savings Plans cost on a daily (24 hourly) basis.

savingsPlansPurchaseRecommendationSummary_estimatedTotalCost :: Lens' SavingsPlansPurchaseRecommendationSummary (Maybe Text)

The estimated total cost of the usage after purchasing the recommended Savings Plans. This is a sum of the cost of Savings Plans during this term, and the remaining On-Demand usage.

savingsPlansPurchaseRecommendationSummary_estimatedROI :: Lens' SavingsPlansPurchaseRecommendationSummary (Maybe Text)

The estimated return on investment that's based on the recommended Savings Plans and estimated savings.

savingsPlansPurchaseRecommendationSummary_estimatedSavingsAmount :: Lens' SavingsPlansPurchaseRecommendationSummary (Maybe Text)

The estimated total savings over the lookback period, based on the purchase of the recommended Savings Plans.

savingsPlansPurchaseRecommendationSummary_estimatedMonthlySavingsAmount :: Lens' SavingsPlansPurchaseRecommendationSummary (Maybe Text)

The estimated monthly savings amount that's based on the recommended Savings Plans purchase.

savingsPlansPurchaseRecommendationSummary_estimatedOnDemandCostWithCurrentCommitment :: Lens' SavingsPlansPurchaseRecommendationSummary (Maybe Text)

The estimated On-Demand costs you would expect with no additional commitment. It's based on your usage of the selected time period and the Savings Plans you own.

savingsPlansPurchaseRecommendationSummary_estimatedSavingsPercentage :: Lens' SavingsPlansPurchaseRecommendationSummary (Maybe Text)

The estimated savings relative to the total cost of On-Demand usage, over the lookback period. This is calculated as estimatedSavingsAmount/ CurrentOnDemandSpend*100.

savingsPlansPurchaseRecommendationSummary_totalRecommendationCount :: Lens' SavingsPlansPurchaseRecommendationSummary (Maybe Text)

The aggregate number of Savings Plans recommendations that exist for your account.

savingsPlansPurchaseRecommendationSummary_currentOnDemandSpend :: Lens' SavingsPlansPurchaseRecommendationSummary (Maybe Text)

The current total on demand spend of the applicable usage types over the lookback period.

savingsPlansPurchaseRecommendationSummary_hourlyCommitmentToPurchase :: Lens' SavingsPlansPurchaseRecommendationSummary (Maybe Text)

The recommended hourly commitment that's based on the recommendation parameters.