Network.AWS.CostExplorer.Types.SavingsPlansUtilization

Description

Documentation

data SavingsPlansUtilization

The measurement of how well you're using your existing Savings Plans.

See: newSavingsPlansUtilization smart constructor.

Constructors

SavingsPlansUtilization'

Fields

unusedCommitment :: Maybe Text

The amount of your Savings Plans commitment that wasn't consumed from Savings Plans eligible usage in a specific period.

usedCommitment :: Maybe Text

The amount of your Savings Plans commitment that was consumed from Savings Plans eligible usage in a specific period.

totalCommitment :: Maybe Text

The total amount of Savings Plans commitment that's been purchased in an account (or set of accounts).

utilizationPercentage :: Maybe Text

The amount of UsedCommitment divided by the TotalCommitment for your Savings Plans.

newSavingsPlansUtilization :: SavingsPlansUtilization

Create a value of SavingsPlansUtilization with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:unusedCommitment:SavingsPlansUtilization', savingsPlansUtilization_unusedCommitment - The amount of your Savings Plans commitment that wasn't consumed from Savings Plans eligible usage in a specific period.

$sel:usedCommitment:SavingsPlansUtilization', savingsPlansUtilization_usedCommitment - The amount of your Savings Plans commitment that was consumed from Savings Plans eligible usage in a specific period.

$sel:totalCommitment:SavingsPlansUtilization', savingsPlansUtilization_totalCommitment - The total amount of Savings Plans commitment that's been purchased in an account (or set of accounts).

$sel:utilizationPercentage:SavingsPlansUtilization', savingsPlansUtilization_utilizationPercentage - The amount of UsedCommitment divided by the TotalCommitment for your Savings Plans.

savingsPlansUtilization_unusedCommitment :: Lens' SavingsPlansUtilization (Maybe Text)

The amount of your Savings Plans commitment that wasn't consumed from Savings Plans eligible usage in a specific period.

savingsPlansUtilization_usedCommitment :: Lens' SavingsPlansUtilization (Maybe Text)

The amount of your Savings Plans commitment that was consumed from Savings Plans eligible usage in a specific period.

savingsPlansUtilization_totalCommitment :: Lens' SavingsPlansUtilization (Maybe Text)

The total amount of Savings Plans commitment that's been purchased in an account (or set of accounts).

savingsPlansUtilization_utilizationPercentage :: Lens' SavingsPlansUtilization (Maybe Text)

The amount of UsedCommitment divided by the TotalCommitment for your Savings Plans.