Network.AWS.CostExplorer.Types.SavingsPlansUtilizationDetail

Description

Documentation

data SavingsPlansUtilizationDetail

A single daily or monthly Savings Plans utilization rate, and details for your account. A management account in an organization have access to member accounts. You can use GetDimensionValues to determine the possible dimension values.

See: newSavingsPlansUtilizationDetail smart constructor.

Constructors

SavingsPlansUtilizationDetail'

Fields

savings :: Maybe SavingsPlansSavings

The amount saved by using existing Savings Plans. Savings returns both net savings from savings plans as well as the onDemandCostEquivalent of the Savings Plans when considering the utilization rate.

utilization :: Maybe SavingsPlansUtilization

A ratio of your effectiveness of using existing Savings Plans to apply to workloads that are Savings Plans eligible.

attributes :: Maybe (HashMap Text Text)

The attribute that applies to a specific Dimension.

amortizedCommitment :: Maybe SavingsPlansAmortizedCommitment

The total amortized commitment for a Savings Plans. Includes the sum of the upfront and recurring Savings Plans fees.

savingsPlanArn :: Maybe Text

The unique Amazon Resource Name (ARN) for a particular Savings Plan.

newSavingsPlansUtilizationDetail :: SavingsPlansUtilizationDetail

Create a value of SavingsPlansUtilizationDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:savings:SavingsPlansUtilizationDetail', savingsPlansUtilizationDetail_savings - The amount saved by using existing Savings Plans. Savings returns both net savings from savings plans as well as the onDemandCostEquivalent of the Savings Plans when considering the utilization rate.

$sel:utilization:SavingsPlansUtilizationDetail', savingsPlansUtilizationDetail_utilization - A ratio of your effectiveness of using existing Savings Plans to apply to workloads that are Savings Plans eligible.

$sel:attributes:SavingsPlansUtilizationDetail', savingsPlansUtilizationDetail_attributes - The attribute that applies to a specific Dimension.

$sel:amortizedCommitment:SavingsPlansUtilizationDetail', savingsPlansUtilizationDetail_amortizedCommitment - The total amortized commitment for a Savings Plans. Includes the sum of the upfront and recurring Savings Plans fees.

$sel:savingsPlanArn:SavingsPlansUtilizationDetail', savingsPlansUtilizationDetail_savingsPlanArn - The unique Amazon Resource Name (ARN) for a particular Savings Plan.

savingsPlansUtilizationDetail_savings :: Lens' SavingsPlansUtilizationDetail (Maybe SavingsPlansSavings)

The amount saved by using existing Savings Plans. Savings returns both net savings from savings plans as well as the onDemandCostEquivalent of the Savings Plans when considering the utilization rate.

savingsPlansUtilizationDetail_utilization :: Lens' SavingsPlansUtilizationDetail (Maybe SavingsPlansUtilization)

A ratio of your effectiveness of using existing Savings Plans to apply to workloads that are Savings Plans eligible.

savingsPlansUtilizationDetail_attributes :: Lens' SavingsPlansUtilizationDetail (Maybe (HashMap Text Text))

The attribute that applies to a specific Dimension.

savingsPlansUtilizationDetail_amortizedCommitment :: Lens' SavingsPlansUtilizationDetail (Maybe SavingsPlansAmortizedCommitment)

The total amortized commitment for a Savings Plans. Includes the sum of the upfront and recurring Savings Plans fees.

savingsPlansUtilizationDetail_savingsPlanArn :: Lens' SavingsPlansUtilizationDetail (Maybe Text)

The unique Amazon Resource Name (ARN) for a particular Savings Plan.