Network.AWS.CostExplorer.Types.TargetInstance

Description

Documentation

data TargetInstance

Details on recommended instance.

See: newTargetInstance smart constructor.

Constructors

TargetInstance'

Fields

estimatedMonthlySavings :: Maybe Text

The estimated savings that result from modification, on a monthly basis.

currencyCode :: Maybe Text

The currency code that Amazon Web Services used to calculate the costs for this instance.

estimatedMonthlyCost :: Maybe Text

The expected cost to operate this instance type on a monthly basis.

expectedResourceUtilization :: Maybe ResourceUtilization

The expected utilization metrics for target instance type.

platformDifferences :: Maybe [PlatformDifference]

Explains the actions you might need to take in order to successfully migrate your workloads from the current instance type to the recommended instance type.

defaultTargetInstance :: Maybe Bool

Determines whether this recommendation is the defaulted Amazon Web Services recommendation.

resourceDetails :: Maybe ResourceDetails

Details on the target instance type.

newTargetInstance :: TargetInstance

Create a value of TargetInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:estimatedMonthlySavings:TargetInstance', targetInstance_estimatedMonthlySavings - The estimated savings that result from modification, on a monthly basis.

$sel:currencyCode:TargetInstance', targetInstance_currencyCode - The currency code that Amazon Web Services used to calculate the costs for this instance.

$sel:estimatedMonthlyCost:TargetInstance', targetInstance_estimatedMonthlyCost - The expected cost to operate this instance type on a monthly basis.

$sel:expectedResourceUtilization:TargetInstance', targetInstance_expectedResourceUtilization - The expected utilization metrics for target instance type.

$sel:platformDifferences:TargetInstance', targetInstance_platformDifferences - Explains the actions you might need to take in order to successfully migrate your workloads from the current instance type to the recommended instance type.

$sel:defaultTargetInstance:TargetInstance', targetInstance_defaultTargetInstance - Determines whether this recommendation is the defaulted Amazon Web Services recommendation.

$sel:resourceDetails:TargetInstance', targetInstance_resourceDetails - Details on the target instance type.

targetInstance_estimatedMonthlySavings :: Lens' TargetInstance (Maybe Text)

The estimated savings that result from modification, on a monthly basis.

targetInstance_currencyCode :: Lens' TargetInstance (Maybe Text)

The currency code that Amazon Web Services used to calculate the costs for this instance.

targetInstance_estimatedMonthlyCost :: Lens' TargetInstance (Maybe Text)

The expected cost to operate this instance type on a monthly basis.

targetInstance_expectedResourceUtilization :: Lens' TargetInstance (Maybe ResourceUtilization)

The expected utilization metrics for target instance type.

targetInstance_platformDifferences :: Lens' TargetInstance (Maybe [PlatformDifference])

Explains the actions you might need to take in order to successfully migrate your workloads from the current instance type to the recommended instance type.

targetInstance_defaultTargetInstance :: Lens' TargetInstance (Maybe Bool)

Determines whether this recommendation is the defaulted Amazon Web Services recommendation.

targetInstance_resourceDetails :: Lens' TargetInstance (Maybe ResourceDetails)

Details on the target instance type.