Network.AWS.CostExplorer.UpdateAnomalySubscription

Description

Updates an existing cost anomaly monitor subscription.

Creating a Request

data UpdateAnomalySubscription

See: newUpdateAnomalySubscription smart constructor.

Constructors

UpdateAnomalySubscription'

Fields

threshold :: Maybe Double

The update to the threshold value for receiving notifications.

subscriptionName :: Maybe Text

The new name of the subscription.

subscribers :: Maybe [Subscriber]

The update to the subscriber list.

frequency :: Maybe AnomalySubscriptionFrequency

The update to the frequency value that subscribers receive notifications.

monitorArnList :: Maybe [Text]

A list of cost anomaly monitor ARNs.

subscriptionArn :: Text

A cost anomaly subscription Amazon Resource Name (ARN).

newUpdateAnomalySubscription

Arguments

:: Text
$sel:subscriptionArn:UpdateAnomalySubscription'

-> UpdateAnomalySubscription

Create a value of UpdateAnomalySubscription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:threshold:UpdateAnomalySubscription', updateAnomalySubscription_threshold - The update to the threshold value for receiving notifications.

$sel:subscriptionName:UpdateAnomalySubscription', updateAnomalySubscription_subscriptionName - The new name of the subscription.

$sel:subscribers:UpdateAnomalySubscription', updateAnomalySubscription_subscribers - The update to the subscriber list.

$sel:frequency:UpdateAnomalySubscription', updateAnomalySubscription_frequency - The update to the frequency value that subscribers receive notifications.

$sel:monitorArnList:UpdateAnomalySubscription', updateAnomalySubscription_monitorArnList - A list of cost anomaly monitor ARNs.

$sel:subscriptionArn:UpdateAnomalySubscription', updateAnomalySubscription_subscriptionArn - A cost anomaly subscription Amazon Resource Name (ARN).

Request Lenses

updateAnomalySubscription_threshold :: Lens' UpdateAnomalySubscription (Maybe Double)

The update to the threshold value for receiving notifications.

updateAnomalySubscription_subscriptionName :: Lens' UpdateAnomalySubscription (Maybe Text)

The new name of the subscription.

updateAnomalySubscription_subscribers :: Lens' UpdateAnomalySubscription (Maybe [Subscriber])

The update to the subscriber list.

updateAnomalySubscription_frequency :: Lens' UpdateAnomalySubscription (Maybe AnomalySubscriptionFrequency)

The update to the frequency value that subscribers receive notifications.

updateAnomalySubscription_monitorArnList :: Lens' UpdateAnomalySubscription (Maybe [Text])

A list of cost anomaly monitor ARNs.

updateAnomalySubscription_subscriptionArn :: Lens' UpdateAnomalySubscription Text

A cost anomaly subscription Amazon Resource Name (ARN).

Destructuring the Response

data UpdateAnomalySubscriptionResponse

See: newUpdateAnomalySubscriptionResponse smart constructor.

Constructors

UpdateAnomalySubscriptionResponse'

Fields

httpStatus :: Int

The response's http status code.

subscriptionArn :: Text

A cost anomaly subscription ARN.

newUpdateAnomalySubscriptionResponse

Arguments

:: Int
$sel:httpStatus:UpdateAnomalySubscriptionResponse'

-> Text
$sel:subscriptionArn:UpdateAnomalySubscription'

-> UpdateAnomalySubscriptionResponse

Create a value of UpdateAnomalySubscriptionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateAnomalySubscriptionResponse', updateAnomalySubscriptionResponse_httpStatus - The response's http status code.

$sel:subscriptionArn:UpdateAnomalySubscription', updateAnomalySubscriptionResponse_subscriptionArn - A cost anomaly subscription ARN.

Response Lenses

updateAnomalySubscriptionResponse_httpStatus :: Lens' UpdateAnomalySubscriptionResponse Int

The response's http status code.

updateAnomalySubscriptionResponse_subscriptionArn :: Lens' UpdateAnomalySubscriptionResponse Text

A cost anomaly subscription ARN.