Network.AWS.CostExplorer.UpdateCostCategoryDefinition

Description

Updates an existing Cost Category. Changes made to the Cost Category rules will be used to categorize the current month’s expenses and future expenses. This won’t change categorization for the previous months.

Creating a Request

data UpdateCostCategoryDefinition

See: newUpdateCostCategoryDefinition smart constructor.

Constructors

UpdateCostCategoryDefinition'

Fields

splitChargeRules :: Maybe (NonEmpty CostCategorySplitChargeRule)

The split charge rules used to allocate your charges between your Cost Category values.

defaultValue :: Maybe Text

costCategoryArn :: Text

The unique identifier for your Cost Category.

ruleVersion :: CostCategoryRuleVersion

rules :: NonEmpty CostCategoryRule

The Expression object used to categorize costs. For more information, see CostCategoryRule .

newUpdateCostCategoryDefinition

Arguments

:: Text
$sel:costCategoryArn:UpdateCostCategoryDefinition'

-> CostCategoryRuleVersion
$sel:ruleVersion:UpdateCostCategoryDefinition'

-> NonEmpty CostCategoryRule
$sel:rules:UpdateCostCategoryDefinition'

-> UpdateCostCategoryDefinition

Create a value of UpdateCostCategoryDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:splitChargeRules:UpdateCostCategoryDefinition', updateCostCategoryDefinition_splitChargeRules - The split charge rules used to allocate your charges between your Cost Category values.

$sel:defaultValue:UpdateCostCategoryDefinition', updateCostCategoryDefinition_defaultValue - Undocumented member.

$sel:costCategoryArn:UpdateCostCategoryDefinition', updateCostCategoryDefinition_costCategoryArn - The unique identifier for your Cost Category.

$sel:ruleVersion:UpdateCostCategoryDefinition', updateCostCategoryDefinition_ruleVersion - Undocumented member.

$sel:rules:UpdateCostCategoryDefinition', updateCostCategoryDefinition_rules - The Expression object used to categorize costs. For more information, see CostCategoryRule .

Request Lenses

updateCostCategoryDefinition_splitChargeRules :: Lens' UpdateCostCategoryDefinition (Maybe (NonEmpty CostCategorySplitChargeRule))

The split charge rules used to allocate your charges between your Cost Category values.

updateCostCategoryDefinition_defaultValue :: Lens' UpdateCostCategoryDefinition (Maybe Text)

Undocumented member.

updateCostCategoryDefinition_costCategoryArn :: Lens' UpdateCostCategoryDefinition Text

The unique identifier for your Cost Category.

updateCostCategoryDefinition_ruleVersion :: Lens' UpdateCostCategoryDefinition CostCategoryRuleVersion

Undocumented member.

updateCostCategoryDefinition_rules :: Lens' UpdateCostCategoryDefinition (NonEmpty CostCategoryRule)

The Expression object used to categorize costs. For more information, see CostCategoryRule .

Destructuring the Response

data UpdateCostCategoryDefinitionResponse

See: newUpdateCostCategoryDefinitionResponse smart constructor.

Constructors

UpdateCostCategoryDefinitionResponse'

Fields

costCategoryArn :: Maybe Text

The unique identifier for your Cost Category.

effectiveStart :: Maybe Text

The Cost Category's effective start date.

httpStatus :: Int

The response's http status code.

newUpdateCostCategoryDefinitionResponse

Arguments

:: Int
$sel:httpStatus:UpdateCostCategoryDefinitionResponse'

-> UpdateCostCategoryDefinitionResponse

Create a value of UpdateCostCategoryDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:costCategoryArn:UpdateCostCategoryDefinition', updateCostCategoryDefinitionResponse_costCategoryArn - The unique identifier for your Cost Category.

$sel:effectiveStart:UpdateCostCategoryDefinitionResponse', updateCostCategoryDefinitionResponse_effectiveStart - The Cost Category's effective start date.

$sel:httpStatus:UpdateCostCategoryDefinitionResponse', updateCostCategoryDefinitionResponse_httpStatus - The response's http status code.

Response Lenses

updateCostCategoryDefinitionResponse_costCategoryArn :: Lens' UpdateCostCategoryDefinitionResponse (Maybe Text)

The unique identifier for your Cost Category.

updateCostCategoryDefinitionResponse_effectiveStart :: Lens' UpdateCostCategoryDefinitionResponse (Maybe Text)

The Cost Category's effective start date.

updateCostCategoryDefinitionResponse_httpStatus :: Lens' UpdateCostCategoryDefinitionResponse Int

The response's http status code.