Network.AWS.Crypto

Description

HMAC

type Key = ByteString

hmacSHA1 :: ByteArrayAccess a => Key -> a -> HMAC SHA1

hmacSHA256 :: ByteArrayAccess a => Key -> a -> HMAC SHA256

Hashing

hashSHA256 :: ByteArrayAccess a => a -> Digest SHA256

hashMD5 :: ByteArrayAccess a => a -> Digest MD5

hash :: (ByteArrayAccess ba, HashAlgorithm a) => ba -> Digest a

Hash a strict bytestring into a digest.

sinkSHA256 :: Monad m => ConduitM ByteString o m (Digest SHA256)

Incrementally calculate a SHA256 Digest.

sinkMD5 :: Monad m => ConduitM ByteString o m (Digest MD5)

Incrementally calculate a MD5 Digest.

Re-exported

class HashAlgorithm a

Class representing hashing algorithms.

The interface presented here is update in place and lowlevel. the Hash module takes care of hidding the mutable interface properly.

Minimal complete definition

hashBlockSize, hashDigestSize, hashInternalContextSize, hashInternalInit, hashInternalUpdate, hashInternalFinalize

data Digest a

Represent a digest for a given hash algorithm.

This type is an instance of ByteArrayAccess from package memory. Module Data.ByteArray provides many primitives to work with those values including conversion to other types.

Creating a digest from a bytearray is also possible with function digestFromByteString.

data SHA256

SHA256 cryptographic hash algorithm

data MD5

MD5 cryptographic hash algorithm