Network.AWS.CustomerProfiles.AddProfileKey

Description

Associates a new key value with a specific profile, such as a Contact Trace Record (CTR) ContactId.

A profile object can have a single unique key and any number of additional keys that can be used to identify the profile that it belongs to.

Creating a Request

data AddProfileKey

See: newAddProfileKey smart constructor.

Constructors

AddProfileKey'

Fields

profileId :: Text

The unique identifier of a customer profile.

keyName :: Text

A searchable identifier of a customer profile.

values :: [Text]

A list of key values.

domainName :: Text

The unique name of the domain.

newAddProfileKey

Arguments

:: Text
$sel:profileId:AddProfileKey'

-> Text
$sel:keyName:AddProfileKey'

-> Text
$sel:domainName:AddProfileKey'

-> AddProfileKey

Create a value of AddProfileKey with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:profileId:AddProfileKey', addProfileKey_profileId - The unique identifier of a customer profile.

$sel:keyName:AddProfileKey', addProfileKey_keyName - A searchable identifier of a customer profile.

$sel:values:AddProfileKey', addProfileKey_values - A list of key values.

$sel:domainName:AddProfileKey', addProfileKey_domainName - The unique name of the domain.

Request Lenses

addProfileKey_profileId :: Lens' AddProfileKey Text

The unique identifier of a customer profile.

addProfileKey_keyName :: Lens' AddProfileKey Text

A searchable identifier of a customer profile.

addProfileKey_values :: Lens' AddProfileKey [Text]

A list of key values.

addProfileKey_domainName :: Lens' AddProfileKey Text

The unique name of the domain.

Destructuring the Response

data AddProfileKeyResponse

See: newAddProfileKeyResponse smart constructor.

Constructors

AddProfileKeyResponse'

Fields

values :: Maybe [Text]

A list of key values.

keyName :: Maybe Text

A searchable identifier of a customer profile.

httpStatus :: Int

The response's http status code.

newAddProfileKeyResponse

Arguments

:: Int
$sel:httpStatus:AddProfileKeyResponse'

-> AddProfileKeyResponse

Create a value of AddProfileKeyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:values:AddProfileKey', addProfileKeyResponse_values - A list of key values.

$sel:keyName:AddProfileKey', addProfileKeyResponse_keyName - A searchable identifier of a customer profile.

$sel:httpStatus:AddProfileKeyResponse', addProfileKeyResponse_httpStatus - The response's http status code.

Response Lenses

addProfileKeyResponse_values :: Lens' AddProfileKeyResponse (Maybe [Text])

A list of key values.

addProfileKeyResponse_keyName :: Lens' AddProfileKeyResponse (Maybe Text)

A searchable identifier of a customer profile.

addProfileKeyResponse_httpStatus :: Lens' AddProfileKeyResponse Int

The response's http status code.