Network.AWS.CustomerProfiles.Lens

Description

Operations

DeleteProfileObjectType

deleteProfileObjectType_domainName :: Lens' DeleteProfileObjectType Text

The unique name of the domain.

deleteProfileObjectType_objectTypeName :: Lens' DeleteProfileObjectType Text

The name of the profile object type.

deleteProfileObjectTypeResponse_httpStatus :: Lens' DeleteProfileObjectTypeResponse Int

The response's http status code.

deleteProfileObjectTypeResponse_message :: Lens' DeleteProfileObjectTypeResponse Text

A message that indicates the delete request is done.

ListIntegrations

listIntegrations_nextToken :: Lens' ListIntegrations (Maybe Text)

The pagination token from the previous ListIntegrations API call.

listIntegrations_maxResults :: Lens' ListIntegrations (Maybe Natural)

The maximum number of objects returned per page.

listIntegrations_domainName :: Lens' ListIntegrations Text

The unique name of the domain.

listIntegrationsResponse_items :: Lens' ListIntegrationsResponse (Maybe [ListIntegrationItem])

The list of ListIntegrations instances.

listIntegrationsResponse_nextToken :: Lens' ListIntegrationsResponse (Maybe Text)

The pagination token from the previous ListIntegrations API call.

listIntegrationsResponse_httpStatus :: Lens' ListIntegrationsResponse Int

The response's http status code.

PutProfileObjectType

putProfileObjectType_expirationDays :: Lens' PutProfileObjectType (Maybe Natural)

The number of days until the data in the object expires.

putProfileObjectType_templateId :: Lens' PutProfileObjectType (Maybe Text)

A unique identifier for the object template.

putProfileObjectType_keys :: Lens' PutProfileObjectType (Maybe (HashMap Text [ObjectTypeKey]))

A list of unique keys that can be used to map data to the profile.

putProfileObjectType_encryptionKey :: Lens' PutProfileObjectType (Maybe Text)

The customer-provided key to encrypt the profile object that will be created in this profile object type.

putProfileObjectType_allowProfileCreation :: Lens' PutProfileObjectType (Maybe Bool)

Indicates whether a profile should be created when data is received if one doesn’t exist for an object of this type. The default is FALSE. If the AllowProfileCreation flag is set to FALSE, then the service tries to fetch a standard profile and associate this object with the profile. If it is set to TRUE, and if no match is found, then the service creates a new standard profile.

putProfileObjectType_fields :: Lens' PutProfileObjectType (Maybe (HashMap Text ObjectTypeField))

A map of the name and ObjectType field.

putProfileObjectType_tags :: Lens' PutProfileObjectType (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

putProfileObjectType_domainName :: Lens' PutProfileObjectType Text

The unique name of the domain.

putProfileObjectType_objectTypeName :: Lens' PutProfileObjectType Text

The name of the profile object type.

putProfileObjectType_description :: Lens' PutProfileObjectType Text

Description of the profile object type.

putProfileObjectTypeResponse_expirationDays :: Lens' PutProfileObjectTypeResponse (Maybe Natural)

The number of days until the data in the object expires.

putProfileObjectTypeResponse_lastUpdatedAt :: Lens' PutProfileObjectTypeResponse (Maybe UTCTime)

The timestamp of when the domain was most recently edited.

putProfileObjectTypeResponse_createdAt :: Lens' PutProfileObjectTypeResponse (Maybe UTCTime)

The timestamp of when the domain was created.

putProfileObjectTypeResponse_templateId :: Lens' PutProfileObjectTypeResponse (Maybe Text)

A unique identifier for the object template.

putProfileObjectTypeResponse_keys :: Lens' PutProfileObjectTypeResponse (Maybe (HashMap Text [ObjectTypeKey]))

A list of unique keys that can be used to map data to the profile.

putProfileObjectTypeResponse_encryptionKey :: Lens' PutProfileObjectTypeResponse (Maybe Text)

The customer-provided key to encrypt the profile object that will be created in this profile object type.

putProfileObjectTypeResponse_allowProfileCreation :: Lens' PutProfileObjectTypeResponse (Maybe Bool)

Indicates whether a profile should be created when data is received if one doesn’t exist for an object of this type. The default is FALSE. If the AllowProfileCreation flag is set to FALSE, then the service tries to fetch a standard profile and associate this object with the profile. If it is set to TRUE, and if no match is found, then the service creates a new standard profile.

putProfileObjectTypeResponse_fields :: Lens' PutProfileObjectTypeResponse (Maybe (HashMap Text ObjectTypeField))

A map of the name and ObjectType field.

putProfileObjectTypeResponse_tags :: Lens' PutProfileObjectTypeResponse (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

putProfileObjectTypeResponse_httpStatus :: Lens' PutProfileObjectTypeResponse Int

The response's http status code.

putProfileObjectTypeResponse_objectTypeName :: Lens' PutProfileObjectTypeResponse Text

The name of the profile object type.

putProfileObjectTypeResponse_description :: Lens' PutProfileObjectTypeResponse Text

Description of the profile object type.

ListProfileObjects

listProfileObjects_objectFilter :: Lens' ListProfileObjects (Maybe ObjectFilter)

Applies a filter to the response to include profile objects with the specified index values. This filter is only supported for ObjectTypeName _asset and _case.

listProfileObjects_nextToken :: Lens' ListProfileObjects (Maybe Text)

The pagination token from the previous call to ListProfileObjects.

listProfileObjects_maxResults :: Lens' ListProfileObjects (Maybe Natural)

The maximum number of objects returned per page.

listProfileObjects_domainName :: Lens' ListProfileObjects Text

The unique name of the domain.

listProfileObjects_objectTypeName :: Lens' ListProfileObjects Text

The name of the profile object type.

listProfileObjects_profileId :: Lens' ListProfileObjects Text

The unique identifier of a customer profile.

listProfileObjectsResponse_items :: Lens' ListProfileObjectsResponse (Maybe [ListProfileObjectsItem])

The list of ListProfileObject instances.

listProfileObjectsResponse_nextToken :: Lens' ListProfileObjectsResponse (Maybe Text)

The pagination token from the previous call to ListProfileObjects.

listProfileObjectsResponse_httpStatus :: Lens' ListProfileObjectsResponse Int

The response's http status code.

ListProfileObjectTypeTemplates

listProfileObjectTypeTemplates_nextToken :: Lens' ListProfileObjectTypeTemplates (Maybe Text)

The pagination token from the previous ListObjectTypeTemplates API call.

listProfileObjectTypeTemplates_maxResults :: Lens' ListProfileObjectTypeTemplates (Maybe Natural)

The maximum number of objects returned per page.

listProfileObjectTypeTemplatesResponse_items :: Lens' ListProfileObjectTypeTemplatesResponse (Maybe [ListProfileObjectTypeTemplateItem])

The list of ListProfileObjectType template instances.

listProfileObjectTypeTemplatesResponse_nextToken :: Lens' ListProfileObjectTypeTemplatesResponse (Maybe Text)

The pagination token from the previous ListObjectTypeTemplates API call.

listProfileObjectTypeTemplatesResponse_httpStatus :: Lens' ListProfileObjectTypeTemplatesResponse Int

The response's http status code.

DeleteProfile

deleteProfile_profileId :: Lens' DeleteProfile Text

The unique identifier of a customer profile.

deleteProfile_domainName :: Lens' DeleteProfile Text

The unique name of the domain.

deleteProfileResponse_message :: Lens' DeleteProfileResponse (Maybe Text)

A message that indicates the delete request is done.

deleteProfileResponse_httpStatus :: Lens' DeleteProfileResponse Int

The response's http status code.

UpdateProfile

updateProfile_shippingAddress :: Lens' UpdateProfile (Maybe UpdateAddress)

The customer’s shipping address.

updateProfile_mobilePhoneNumber :: Lens' UpdateProfile (Maybe Text)

The customer’s mobile phone number.

updateProfile_mailingAddress :: Lens' UpdateProfile (Maybe UpdateAddress)

The customer’s mailing address.

updateProfile_middleName :: Lens' UpdateProfile (Maybe Text)

The customer’s middle name.

updateProfile_personalEmailAddress :: Lens' UpdateProfile (Maybe Text)

The customer’s personal email address.

updateProfile_lastName :: Lens' UpdateProfile (Maybe Text)

The customer’s last name.

updateProfile_additionalInformation :: Lens' UpdateProfile (Maybe Text)

Any additional information relevant to the customer’s profile.

updateProfile_homePhoneNumber :: Lens' UpdateProfile (Maybe Text)

The customer’s home phone number.

updateProfile_address :: Lens' UpdateProfile (Maybe UpdateAddress)

A generic address associated with the customer that is not mailing, shipping, or billing.

updateProfile_partyType :: Lens' UpdateProfile (Maybe PartyType)

The type of profile used to describe the customer.

updateProfile_businessEmailAddress :: Lens' UpdateProfile (Maybe Text)

The customer’s business email address.

updateProfile_attributes :: Lens' UpdateProfile (Maybe (HashMap Text Text))

A key value pair of attributes of a customer profile.

updateProfile_gender :: Lens' UpdateProfile (Maybe Gender)

The gender with which the customer identifies.

updateProfile_phoneNumber :: Lens' UpdateProfile (Maybe Text)

The customer’s phone number, which has not been specified as a mobile, home, or business number.

updateProfile_accountNumber :: Lens' UpdateProfile (Maybe Text)

A unique account number that you have given to the customer.

updateProfile_emailAddress :: Lens' UpdateProfile (Maybe Text)

The customer’s email address, which has not been specified as a personal or business address.

updateProfile_firstName :: Lens' UpdateProfile (Maybe Text)

The customer’s first name.

updateProfile_billingAddress :: Lens' UpdateProfile (Maybe UpdateAddress)

The customer’s billing address.

updateProfile_businessPhoneNumber :: Lens' UpdateProfile (Maybe Text)

The customer’s business phone number.

updateProfile_birthDate :: Lens' UpdateProfile (Maybe Text)

The customer’s birth date.

updateProfile_businessName :: Lens' UpdateProfile (Maybe Text)

The name of the customer’s business.

updateProfile_domainName :: Lens' UpdateProfile Text

The unique name of the domain.

updateProfile_profileId :: Lens' UpdateProfile Text

The unique identifier of a customer profile.

updateProfileResponse_httpStatus :: Lens' UpdateProfileResponse Int

The response's http status code.

updateProfileResponse_profileId :: Lens' UpdateProfileResponse Text

The unique identifier of a customer profile.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The ARN of the resource for which you want to view tags.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

AddProfileKey

addProfileKey_profileId :: Lens' AddProfileKey Text

The unique identifier of a customer profile.

addProfileKey_keyName :: Lens' AddProfileKey Text

A searchable identifier of a customer profile.

addProfileKey_values :: Lens' AddProfileKey [Text]

A list of key values.

addProfileKey_domainName :: Lens' AddProfileKey Text

The unique name of the domain.

addProfileKeyResponse_values :: Lens' AddProfileKeyResponse (Maybe [Text])

A list of key values.

addProfileKeyResponse_keyName :: Lens' AddProfileKeyResponse (Maybe Text)

A searchable identifier of a customer profile.

addProfileKeyResponse_httpStatus :: Lens' AddProfileKeyResponse Int

The response's http status code.

GetProfileObjectTypeTemplate

getProfileObjectTypeTemplate_templateId :: Lens' GetProfileObjectTypeTemplate Text

A unique identifier for the object template.

getProfileObjectTypeTemplateResponse_sourceName :: Lens' GetProfileObjectTypeTemplateResponse (Maybe Text)

The name of the source of the object template.

getProfileObjectTypeTemplateResponse_templateId :: Lens' GetProfileObjectTypeTemplateResponse (Maybe Text)

A unique identifier for the object template.

getProfileObjectTypeTemplateResponse_sourceObject :: Lens' GetProfileObjectTypeTemplateResponse (Maybe Text)

The source of the object template.

getProfileObjectTypeTemplateResponse_keys :: Lens' GetProfileObjectTypeTemplateResponse (Maybe (HashMap Text [ObjectTypeKey]))

A list of unique keys that can be used to map data to the profile.

getProfileObjectTypeTemplateResponse_allowProfileCreation :: Lens' GetProfileObjectTypeTemplateResponse (Maybe Bool)

Indicates whether a profile should be created when data is received if one doesn’t exist for an object of this type. The default is FALSE. If the AllowProfileCreation flag is set to FALSE, then the service tries to fetch a standard profile and associate this object with the profile. If it is set to TRUE, and if no match is found, then the service creates a new standard profile.

getProfileObjectTypeTemplateResponse_fields :: Lens' GetProfileObjectTypeTemplateResponse (Maybe (HashMap Text ObjectTypeField))

A map of the name and ObjectType field.

getProfileObjectTypeTemplateResponse_httpStatus :: Lens' GetProfileObjectTypeTemplateResponse Int

The response's http status code.

GetIntegration

getIntegration_domainName :: Lens' GetIntegration Text

The unique name of the domain.

getIntegration_uri :: Lens' GetIntegration Text

The URI of the S3 bucket or any other type of data source.

getIntegrationResponse_tags :: Lens' GetIntegrationResponse (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

getIntegrationResponse_httpStatus :: Lens' GetIntegrationResponse Int

The response's http status code.

getIntegrationResponse_domainName :: Lens' GetIntegrationResponse Text

The unique name of the domain.

getIntegrationResponse_uri :: Lens' GetIntegrationResponse Text

The URI of the S3 bucket or any other type of data source.

getIntegrationResponse_objectTypeName :: Lens' GetIntegrationResponse Text

The name of the profile object type.

getIntegrationResponse_createdAt :: Lens' GetIntegrationResponse UTCTime

The timestamp of when the domain was created.

getIntegrationResponse_lastUpdatedAt :: Lens' GetIntegrationResponse UTCTime

The timestamp of when the domain was most recently edited.

GetDomain

getDomain_domainName :: Lens' GetDomain Text

The unique name of the domain.

getDomainResponse_defaultExpirationDays :: Lens' GetDomainResponse (Maybe Natural)

The default number of days until the data within the domain expires.

getDomainResponse_defaultEncryptionKey :: Lens' GetDomainResponse (Maybe Text)

The default encryption key, which is an AWS managed key, is used when no specific type of encryption key is specified. It is used to encrypt all data before it is placed in permanent or semi-permanent storage.

getDomainResponse_matching :: Lens' GetDomainResponse (Maybe MatchingResponse)

The process of matching duplicate profiles. If Matching = true, Amazon Connect Customer Profiles starts a weekly batch process every Saturday at 12AM UTC to detect duplicate profiles in your domains. After that batch process completes, use the GetMatches API to return and review the results.

getDomainResponse_stats :: Lens' GetDomainResponse (Maybe DomainStats)

Usage-specific statistics about the domain.

getDomainResponse_deadLetterQueueUrl :: Lens' GetDomainResponse (Maybe Text)

The URL of the SQS dead letter queue, which is used for reporting errors associated with ingesting data from third party applications.

getDomainResponse_tags :: Lens' GetDomainResponse (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

getDomainResponse_httpStatus :: Lens' GetDomainResponse Int

The response's http status code.

getDomainResponse_domainName :: Lens' GetDomainResponse Text

The unique name of the domain.

getDomainResponse_createdAt :: Lens' GetDomainResponse UTCTime

The timestamp of when the domain was created.

getDomainResponse_lastUpdatedAt :: Lens' GetDomainResponse UTCTime

The timestamp of when the domain was most recently edited.

CreateDomain

createDomain_defaultEncryptionKey :: Lens' CreateDomain (Maybe Text)

The default encryption key, which is an AWS managed key, is used when no specific type of encryption key is specified. It is used to encrypt all data before it is placed in permanent or semi-permanent storage.

createDomain_matching :: Lens' CreateDomain (Maybe MatchingRequest)

The process of matching duplicate profiles. If Matching = true, Amazon Connect Customer Profiles starts a weekly batch process every Saturday at 12AM UTC to detect duplicate profiles in your domains. After that batch process completes, use the GetMatches API to return and review the results.

createDomain_deadLetterQueueUrl :: Lens' CreateDomain (Maybe Text)

The URL of the SQS dead letter queue, which is used for reporting errors associated with ingesting data from third party applications. You must set up a policy on the DeadLetterQueue for the SendMessage operation to enable Amazon Connect Customer Profiles to send messages to the DeadLetterQueue.

createDomain_tags :: Lens' CreateDomain (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

createDomain_domainName :: Lens' CreateDomain Text

The unique name of the domain.

createDomain_defaultExpirationDays :: Lens' CreateDomain Natural

The default number of days until the data within the domain expires.

createDomainResponse_defaultEncryptionKey :: Lens' CreateDomainResponse (Maybe Text)

The default encryption key, which is an AWS managed key, is used when no specific type of encryption key is specified. It is used to encrypt all data before it is placed in permanent or semi-permanent storage.

createDomainResponse_matching :: Lens' CreateDomainResponse (Maybe MatchingResponse)

The process of matching duplicate profiles. If Matching = true, Amazon Connect Customer Profiles starts a weekly batch process every Saturday at 12AM UTC to detect duplicate profiles in your domains. After that batch process completes, use the GetMatches API to return and review the results.

createDomainResponse_deadLetterQueueUrl :: Lens' CreateDomainResponse (Maybe Text)

The URL of the SQS dead letter queue, which is used for reporting errors associated with ingesting data from third party applications.

createDomainResponse_tags :: Lens' CreateDomainResponse (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

createDomainResponse_httpStatus :: Lens' CreateDomainResponse Int

The response's http status code.

createDomainResponse_domainName :: Lens' CreateDomainResponse Text

The unique name of the domain.

createDomainResponse_defaultExpirationDays :: Lens' CreateDomainResponse Natural

The default number of days until the data within the domain expires.

createDomainResponse_createdAt :: Lens' CreateDomainResponse UTCTime

The timestamp of when the domain was created.

createDomainResponse_lastUpdatedAt :: Lens' CreateDomainResponse UTCTime

The timestamp of when the domain was most recently edited.

DeleteIntegration

deleteIntegration_domainName :: Lens' DeleteIntegration Text

The unique name of the domain.

deleteIntegration_uri :: Lens' DeleteIntegration Text

The URI of the S3 bucket or any other type of data source.

deleteIntegrationResponse_httpStatus :: Lens' DeleteIntegrationResponse Int

The response's http status code.

deleteIntegrationResponse_message :: Lens' DeleteIntegrationResponse Text

A message that indicates the delete request is done.

CreateProfile

createProfile_shippingAddress :: Lens' CreateProfile (Maybe Address)

The customer’s shipping address.

createProfile_mobilePhoneNumber :: Lens' CreateProfile (Maybe Text)

The customer’s mobile phone number.

createProfile_mailingAddress :: Lens' CreateProfile (Maybe Address)

The customer’s mailing address.

createProfile_middleName :: Lens' CreateProfile (Maybe Text)

The customer’s middle name.

createProfile_personalEmailAddress :: Lens' CreateProfile (Maybe Text)

The customer’s personal email address.

createProfile_lastName :: Lens' CreateProfile (Maybe Text)

The customer’s last name.

createProfile_additionalInformation :: Lens' CreateProfile (Maybe Text)

Any additional information relevant to the customer’s profile.

createProfile_homePhoneNumber :: Lens' CreateProfile (Maybe Text)

The customer’s home phone number.

createProfile_address :: Lens' CreateProfile (Maybe Address)

A generic address associated with the customer that is not mailing, shipping, or billing.

createProfile_partyType :: Lens' CreateProfile (Maybe PartyType)

The type of profile used to describe the customer.

createProfile_businessEmailAddress :: Lens' CreateProfile (Maybe Text)

The customer’s business email address.

createProfile_attributes :: Lens' CreateProfile (Maybe (HashMap Text Text))

A key value pair of attributes of a customer profile.

createProfile_gender :: Lens' CreateProfile (Maybe Gender)

The gender with which the customer identifies.

createProfile_phoneNumber :: Lens' CreateProfile (Maybe Text)

The customer’s phone number, which has not been specified as a mobile, home, or business number.

createProfile_accountNumber :: Lens' CreateProfile (Maybe Text)

A unique account number that you have given to the customer.

createProfile_emailAddress :: Lens' CreateProfile (Maybe Text)

The customer’s email address, which has not been specified as a personal or business address.

createProfile_firstName :: Lens' CreateProfile (Maybe Text)

The customer’s first name.

createProfile_billingAddress :: Lens' CreateProfile (Maybe Address)

The customer’s billing address.

createProfile_businessPhoneNumber :: Lens' CreateProfile (Maybe Text)

The customer’s business phone number.

createProfile_birthDate :: Lens' CreateProfile (Maybe Text)

The customer’s birth date.

createProfile_businessName :: Lens' CreateProfile (Maybe Text)

The name of the customer’s business.

createProfile_domainName :: Lens' CreateProfile Text

The unique name of the domain.

createProfileResponse_httpStatus :: Lens' CreateProfileResponse Int

The response's http status code.

createProfileResponse_profileId :: Lens' CreateProfileResponse Text

The unique identifier of a customer profile.

PutProfileObject

putProfileObject_objectTypeName :: Lens' PutProfileObject Text

The name of the profile object type.

putProfileObject_object :: Lens' PutProfileObject Text

A string that is serialized from a JSON object.

putProfileObject_domainName :: Lens' PutProfileObject Text

The unique name of the domain.

putProfileObjectResponse_profileObjectUniqueKey :: Lens' PutProfileObjectResponse (Maybe Text)

The unique identifier of the profile object generated by the service.

putProfileObjectResponse_httpStatus :: Lens' PutProfileObjectResponse Int

The response's http status code.

PutIntegration

putIntegration_flowDefinition :: Lens' PutIntegration (Maybe FlowDefinition)

The configuration that controls how Customer Profiles retrieves data from the source.

putIntegration_uri :: Lens' PutIntegration (Maybe Text)

The URI of the S3 bucket or any other type of data source.

putIntegration_tags :: Lens' PutIntegration (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

putIntegration_domainName :: Lens' PutIntegration Text

The unique name of the domain.

putIntegration_objectTypeName :: Lens' PutIntegration Text

The name of the profile object type.

putIntegrationResponse_tags :: Lens' PutIntegrationResponse (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

putIntegrationResponse_httpStatus :: Lens' PutIntegrationResponse Int

The response's http status code.

putIntegrationResponse_domainName :: Lens' PutIntegrationResponse Text

The unique name of the domain.

putIntegrationResponse_uri :: Lens' PutIntegrationResponse Text

The URI of the S3 bucket or any other type of data source.

putIntegrationResponse_objectTypeName :: Lens' PutIntegrationResponse Text

The name of the profile object type.

putIntegrationResponse_createdAt :: Lens' PutIntegrationResponse UTCTime

The timestamp of when the domain was created.

putIntegrationResponse_lastUpdatedAt :: Lens' PutIntegrationResponse UTCTime

The timestamp of when the domain was most recently edited.

DeleteProfileObject

deleteProfileObject_profileId :: Lens' DeleteProfileObject Text

The unique identifier of a customer profile.

deleteProfileObject_profileObjectUniqueKey :: Lens' DeleteProfileObject Text

The unique identifier of the profile object generated by the service.

deleteProfileObject_objectTypeName :: Lens' DeleteProfileObject Text

The name of the profile object type.

deleteProfileObject_domainName :: Lens' DeleteProfileObject Text

The unique name of the domain.

deleteProfileObjectResponse_message :: Lens' DeleteProfileObjectResponse (Maybe Text)

A message that indicates the delete request is done.

deleteProfileObjectResponse_httpStatus :: Lens' DeleteProfileObjectResponse Int

The response's http status code.

ListProfileObjectTypes

listProfileObjectTypes_nextToken :: Lens' ListProfileObjectTypes (Maybe Text)

Identifies the next page of results to return.

listProfileObjectTypes_maxResults :: Lens' ListProfileObjectTypes (Maybe Natural)

The maximum number of objects returned per page.

listProfileObjectTypes_domainName :: Lens' ListProfileObjectTypes Text

The unique name of the domain.

listProfileObjectTypesResponse_items :: Lens' ListProfileObjectTypesResponse (Maybe [ListProfileObjectTypeItem])

The list of ListProfileObjectTypes instances.

listProfileObjectTypesResponse_nextToken :: Lens' ListProfileObjectTypesResponse (Maybe Text)

Identifies the next page of results to return.

listProfileObjectTypesResponse_httpStatus :: Lens' ListProfileObjectTypesResponse Int

The response's http status code.

DeleteProfileKey

deleteProfileKey_profileId :: Lens' DeleteProfileKey Text

The unique identifier of a customer profile.

deleteProfileKey_keyName :: Lens' DeleteProfileKey Text

A searchable identifier of a customer profile.

deleteProfileKey_values :: Lens' DeleteProfileKey [Text]

A list of key values.

deleteProfileKey_domainName :: Lens' DeleteProfileKey Text

The unique name of the domain.

deleteProfileKeyResponse_message :: Lens' DeleteProfileKeyResponse (Maybe Text)

A message that indicates the delete request is done.

deleteProfileKeyResponse_httpStatus :: Lens' DeleteProfileKeyResponse Int

The response's http status code.

GetProfileObjectType

getProfileObjectType_domainName :: Lens' GetProfileObjectType Text

The unique name of the domain.

getProfileObjectType_objectTypeName :: Lens' GetProfileObjectType Text

The name of the profile object type.

getProfileObjectTypeResponse_expirationDays :: Lens' GetProfileObjectTypeResponse (Maybe Natural)

The number of days until the data in the object expires.

getProfileObjectTypeResponse_lastUpdatedAt :: Lens' GetProfileObjectTypeResponse (Maybe UTCTime)

The timestamp of when the domain was most recently edited.

getProfileObjectTypeResponse_createdAt :: Lens' GetProfileObjectTypeResponse (Maybe UTCTime)

The timestamp of when the domain was created.

getProfileObjectTypeResponse_templateId :: Lens' GetProfileObjectTypeResponse (Maybe Text)

A unique identifier for the object template.

getProfileObjectTypeResponse_keys :: Lens' GetProfileObjectTypeResponse (Maybe (HashMap Text [ObjectTypeKey]))

A list of unique keys that can be used to map data to the profile.

getProfileObjectTypeResponse_encryptionKey :: Lens' GetProfileObjectTypeResponse (Maybe Text)

The customer-provided key to encrypt the profile object that will be created in this profile object type.

getProfileObjectTypeResponse_allowProfileCreation :: Lens' GetProfileObjectTypeResponse (Maybe Bool)

Indicates whether a profile should be created when data is received if one doesn’t exist for an object of this type. The default is FALSE. If the AllowProfileCreation flag is set to FALSE, then the service tries to fetch a standard profile and associate this object with the profile. If it is set to TRUE, and if no match is found, then the service creates a new standard profile.

getProfileObjectTypeResponse_fields :: Lens' GetProfileObjectTypeResponse (Maybe (HashMap Text ObjectTypeField))

A map of the name and ObjectType field.

getProfileObjectTypeResponse_tags :: Lens' GetProfileObjectTypeResponse (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

getProfileObjectTypeResponse_httpStatus :: Lens' GetProfileObjectTypeResponse Int

The response's http status code.

getProfileObjectTypeResponse_objectTypeName :: Lens' GetProfileObjectTypeResponse Text

The name of the profile object type.

getProfileObjectTypeResponse_description :: Lens' GetProfileObjectTypeResponse Text

The description of the profile object type.

MergeProfiles

mergeProfiles_fieldSourceProfileIds :: Lens' MergeProfiles (Maybe FieldSourceProfileIds)

The identifiers of the fields in the profile that has the information you want to apply to the merge. For example, say you want to merge EmailAddress from Profile1 into MainProfile. This would be the identifier of the EmailAddress field in Profile1.

mergeProfiles_domainName :: Lens' MergeProfiles Text

The unique name of the domain.

mergeProfiles_mainProfileId :: Lens' MergeProfiles Text

The identifier of the profile to be taken.

mergeProfiles_profileIdsToBeMerged :: Lens' MergeProfiles (NonEmpty Text)

The identifier of the profile to be merged into MainProfileId.

mergeProfilesResponse_message :: Lens' MergeProfilesResponse (Maybe Text)

A message that indicates the merge request is complete.

mergeProfilesResponse_httpStatus :: Lens' MergeProfilesResponse Int

The response's http status code.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The ARN of the resource that you're adding tags to.

tagResource_tags :: Lens' TagResource (HashMap Text Text)

The tags used to organize, track, or control access for this resource.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

GetMatches

getMatches_nextToken :: Lens' GetMatches (Maybe Text)

The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

getMatches_maxResults :: Lens' GetMatches (Maybe Natural)

The maximum number of results to return per page.

getMatches_domainName :: Lens' GetMatches Text

The unique name of the domain.

getMatchesResponse_matches :: Lens' GetMatchesResponse (Maybe [MatchItem])

The list of matched profiles for this instance.

getMatchesResponse_matchGenerationDate :: Lens' GetMatchesResponse (Maybe UTCTime)

The timestamp this version of Match Result generated.

getMatchesResponse_nextToken :: Lens' GetMatchesResponse (Maybe Text)

If there are additional results, this is the token for the next set of results.

getMatchesResponse_potentialMatches :: Lens' GetMatchesResponse (Maybe Natural)

The number of potential matches found.

getMatchesResponse_httpStatus :: Lens' GetMatchesResponse Int

The response's http status code.

UntagResource

untagResource_resourceArn :: Lens' UntagResource Text

The ARN of the resource from which you are removing tags.

untagResource_tagKeys :: Lens' UntagResource (NonEmpty Text)

The list of tag keys to remove from the resource.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

SearchProfiles

searchProfiles_nextToken :: Lens' SearchProfiles (Maybe Text)

The pagination token from the previous SearchProfiles API call.

searchProfiles_maxResults :: Lens' SearchProfiles (Maybe Natural)

The maximum number of objects returned per page.

searchProfiles_domainName :: Lens' SearchProfiles Text

The unique name of the domain.

searchProfiles_keyName :: Lens' SearchProfiles Text

A searchable identifier of a customer profile. The predefined keys you can use to search include: _account, _profileId, _fullName, _phone, _email, _ctrContactId, _marketoLeadId, _salesforceAccountId, _salesforceContactId, _zendeskUserId, _zendeskExternalId, _serviceNowSystemId.

searchProfiles_values :: Lens' SearchProfiles [Text]

A list of key values.

searchProfilesResponse_items :: Lens' SearchProfilesResponse (Maybe [Profile])

The list of SearchProfiles instances.

searchProfilesResponse_nextToken :: Lens' SearchProfilesResponse (Maybe Text)

The pagination token from the previous SearchProfiles API call.

searchProfilesResponse_httpStatus :: Lens' SearchProfilesResponse Int

The response's http status code.

ListAccountIntegrations

listAccountIntegrations_nextToken :: Lens' ListAccountIntegrations (Maybe Text)

The pagination token from the previous ListAccountIntegrations API call.

listAccountIntegrations_maxResults :: Lens' ListAccountIntegrations (Maybe Natural)

The maximum number of objects returned per page.

listAccountIntegrations_uri :: Lens' ListAccountIntegrations Text

The URI of the S3 bucket or any other type of data source.

listAccountIntegrationsResponse_items :: Lens' ListAccountIntegrationsResponse (Maybe [ListIntegrationItem])

The list of ListAccountIntegration instances.

listAccountIntegrationsResponse_nextToken :: Lens' ListAccountIntegrationsResponse (Maybe Text)

The pagination token from the previous ListAccountIntegrations API call.

listAccountIntegrationsResponse_httpStatus :: Lens' ListAccountIntegrationsResponse Int

The response's http status code.

DeleteDomain

deleteDomain_domainName :: Lens' DeleteDomain Text

The unique name of the domain.

deleteDomainResponse_httpStatus :: Lens' DeleteDomainResponse Int

The response's http status code.

deleteDomainResponse_message :: Lens' DeleteDomainResponse Text

A message that indicates the delete request is done.

UpdateDomain

updateDomain_defaultExpirationDays :: Lens' UpdateDomain (Maybe Natural)

The default number of days until the data within the domain expires.

updateDomain_defaultEncryptionKey :: Lens' UpdateDomain (Maybe Text)

The default encryption key, which is an AWS managed key, is used when no specific type of encryption key is specified. It is used to encrypt all data before it is placed in permanent or semi-permanent storage. If specified as an empty string, it will clear any existing value.

updateDomain_matching :: Lens' UpdateDomain (Maybe MatchingRequest)

The process of matching duplicate profiles. If Matching = true, Amazon Connect Customer Profiles starts a weekly batch process every Saturday at 12AM UTC to detect duplicate profiles in your domains. After that batch process completes, use the GetMatches API to return and review the results.

updateDomain_deadLetterQueueUrl :: Lens' UpdateDomain (Maybe Text)

The URL of the SQS dead letter queue, which is used for reporting errors associated with ingesting data from third party applications. If specified as an empty string, it will clear any existing value. You must set up a policy on the DeadLetterQueue for the SendMessage operation to enable Amazon Connect Customer Profiles to send messages to the DeadLetterQueue.

updateDomain_tags :: Lens' UpdateDomain (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

updateDomain_domainName :: Lens' UpdateDomain Text

The unique name of the domain.

updateDomainResponse_defaultExpirationDays :: Lens' UpdateDomainResponse (Maybe Natural)

The default number of days until the data within the domain expires.

updateDomainResponse_defaultEncryptionKey :: Lens' UpdateDomainResponse (Maybe Text)

The default encryption key, which is an AWS managed key, is used when no specific type of encryption key is specified. It is used to encrypt all data before it is placed in permanent or semi-permanent storage.

updateDomainResponse_matching :: Lens' UpdateDomainResponse (Maybe MatchingResponse)

The process of matching duplicate profiles. If Matching = true, Amazon Connect Customer Profiles starts a weekly batch process every Saturday at 12AM UTC to detect duplicate profiles in your domains. After that batch process completes, use the GetMatches API to return and review the results.

updateDomainResponse_deadLetterQueueUrl :: Lens' UpdateDomainResponse (Maybe Text)

The URL of the SQS dead letter queue, which is used for reporting errors associated with ingesting data from third party applications.

updateDomainResponse_tags :: Lens' UpdateDomainResponse (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

updateDomainResponse_httpStatus :: Lens' UpdateDomainResponse Int

The response's http status code.

updateDomainResponse_domainName :: Lens' UpdateDomainResponse Text

The unique name of the domain.

updateDomainResponse_createdAt :: Lens' UpdateDomainResponse UTCTime

The timestamp of when the domain was created.

updateDomainResponse_lastUpdatedAt :: Lens' UpdateDomainResponse UTCTime

The timestamp of when the domain was most recently edited.

ListDomains

listDomains_nextToken :: Lens' ListDomains (Maybe Text)

The pagination token from the previous ListDomain API call.

listDomains_maxResults :: Lens' ListDomains (Maybe Natural)

The maximum number of objects returned per page.

listDomainsResponse_items :: Lens' ListDomainsResponse (Maybe [ListDomainItem])

The list of ListDomains instances.

listDomainsResponse_nextToken :: Lens' ListDomainsResponse (Maybe Text)

The pagination token from the previous ListDomains API call.

listDomainsResponse_httpStatus :: Lens' ListDomainsResponse Int

The response's http status code.

Types

Address

address_address2 :: Lens' Address (Maybe Text)

The second line of a customer address.

address_state :: Lens' Address (Maybe Text)

The state in which a customer lives.

address_address4 :: Lens' Address (Maybe Text)

The fourth line of a customer address.

address_county :: Lens' Address (Maybe Text)

The county in which a customer lives.

address_address3 :: Lens' Address (Maybe Text)

The third line of a customer address.

address_postalCode :: Lens' Address (Maybe Text)

The postal code of a customer address.

address_country :: Lens' Address (Maybe Text)

The country in which a customer lives.

address_city :: Lens' Address (Maybe Text)

The city in which a customer lives.

address_address1 :: Lens' Address (Maybe Text)

The first line of a customer address.

address_province :: Lens' Address (Maybe Text)

The province in which a customer lives.

ConnectorOperator

connectorOperator_serviceNow :: Lens' ConnectorOperator (Maybe ServiceNowConnectorOperator)

The operation to be performed on the provided ServiceNow source fields.

connectorOperator_marketo :: Lens' ConnectorOperator (Maybe MarketoConnectorOperator)

The operation to be performed on the provided Marketo source fields.

connectorOperator_salesforce :: Lens' ConnectorOperator (Maybe SalesforceConnectorOperator)

The operation to be performed on the provided Salesforce source fields.

connectorOperator_zendesk :: Lens' ConnectorOperator (Maybe ZendeskConnectorOperator)

The operation to be performed on the provided Zendesk source fields.

connectorOperator_s3 :: Lens' ConnectorOperator (Maybe S3ConnectorOperator)

The operation to be performed on the provided Amazon S3 source fields.

DomainStats

domainStats_meteringProfileCount :: Lens' DomainStats (Maybe Integer)

The number of profiles that you are currently paying for in the domain. If you have more than 100 objects associated with a single profile, that profile counts as two profiles. If you have more than 200 objects, that profile counts as three, and so on.

domainStats_totalSize :: Lens' DomainStats (Maybe Integer)

The total size, in bytes, of all objects in the domain.

domainStats_profileCount :: Lens' DomainStats (Maybe Integer)

The total number of profiles currently in the domain.

domainStats_objectCount :: Lens' DomainStats (Maybe Integer)

The total number of objects in domain.

FieldSourceProfileIds

fieldSourceProfileIds_shippingAddress :: Lens' FieldSourceProfileIds (Maybe Text)

A unique identifier for the shipping address field to be merged.

fieldSourceProfileIds_mobilePhoneNumber :: Lens' FieldSourceProfileIds (Maybe Text)

A unique identifier for the mobile phone number field to be merged.

fieldSourceProfileIds_mailingAddress :: Lens' FieldSourceProfileIds (Maybe Text)

A unique identifier for the mailing address field to be merged.

fieldSourceProfileIds_middleName :: Lens' FieldSourceProfileIds (Maybe Text)

A unique identifier for the middle name field to be merged.

fieldSourceProfileIds_personalEmailAddress :: Lens' FieldSourceProfileIds (Maybe Text)

A unique identifier for the personal email address field to be merged.

fieldSourceProfileIds_lastName :: Lens' FieldSourceProfileIds (Maybe Text)

A unique identifier for the last name field to be merged.

fieldSourceProfileIds_additionalInformation :: Lens' FieldSourceProfileIds (Maybe Text)

A unique identifier for the additional information field to be merged.

fieldSourceProfileIds_homePhoneNumber :: Lens' FieldSourceProfileIds (Maybe Text)

A unique identifier for the home phone number field to be merged.

fieldSourceProfileIds_address :: Lens' FieldSourceProfileIds (Maybe Text)

A unique identifier for the party type field to be merged.

fieldSourceProfileIds_partyType :: Lens' FieldSourceProfileIds (Maybe Text)

A unique identifier for the party type field to be merged.

fieldSourceProfileIds_businessEmailAddress :: Lens' FieldSourceProfileIds (Maybe Text)

A unique identifier for the party type field to be merged.

fieldSourceProfileIds_attributes :: Lens' FieldSourceProfileIds (Maybe (HashMap Text Text))

A unique identifier for the attributes field to be merged.

fieldSourceProfileIds_gender :: Lens' FieldSourceProfileIds (Maybe Text)

A unique identifier for the gender field to be merged.

fieldSourceProfileIds_phoneNumber :: Lens' FieldSourceProfileIds (Maybe Text)

A unique identifier for the phone number field to be merged.

fieldSourceProfileIds_accountNumber :: Lens' FieldSourceProfileIds (Maybe Text)

A unique identifier for the account number field to be merged.

fieldSourceProfileIds_emailAddress :: Lens' FieldSourceProfileIds (Maybe Text)

A unique identifier for the email address field to be merged.

fieldSourceProfileIds_firstName :: Lens' FieldSourceProfileIds (Maybe Text)

A unique identifier for the first name field to be merged.

fieldSourceProfileIds_billingAddress :: Lens' FieldSourceProfileIds (Maybe Text)

A unique identifier for the billing type field to be merged.

fieldSourceProfileIds_businessPhoneNumber :: Lens' FieldSourceProfileIds (Maybe Text)

A unique identifier for the business phone number field to be merged.

fieldSourceProfileIds_birthDate :: Lens' FieldSourceProfileIds (Maybe Text)

A unique identifier for the birthdate field to be merged.

fieldSourceProfileIds_businessName :: Lens' FieldSourceProfileIds (Maybe Text)

A unique identifier for the business name field to be merged.

FlowDefinition

flowDefinition_description :: Lens' FlowDefinition (Maybe Text)

A description of the flow you want to create.

flowDefinition_flowName :: Lens' FlowDefinition Text

The specified name of the flow. Use underscores (_) or hyphens (-) only. Spaces are not allowed.

flowDefinition_kmsArn :: Lens' FlowDefinition Text

The Amazon Resource Name of the AWS Key Management Service (KMS) key you provide for encryption.

flowDefinition_sourceFlowConfig :: Lens' FlowDefinition SourceFlowConfig

The configuration that controls how Customer Profiles retrieves data from the source.

flowDefinition_tasks :: Lens' FlowDefinition [Task]

A list of tasks that Customer Profiles performs while transferring the data in the flow run.

flowDefinition_triggerConfig :: Lens' FlowDefinition TriggerConfig

The trigger settings that determine how and when the flow runs.

IncrementalPullConfig

incrementalPullConfig_datetimeTypeFieldName :: Lens' IncrementalPullConfig (Maybe Text)

A field that specifies the date time or timestamp field as the criteria to use when importing incremental records from the source.

ListDomainItem

listDomainItem_tags :: Lens' ListDomainItem (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

listDomainItem_domainName :: Lens' ListDomainItem Text

The unique name of the domain.

listDomainItem_createdAt :: Lens' ListDomainItem UTCTime

The timestamp of when the domain was created.

listDomainItem_lastUpdatedAt :: Lens' ListDomainItem UTCTime

The timestamp of when the domain was most recently edited.

ListIntegrationItem

listIntegrationItem_tags :: Lens' ListIntegrationItem (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

listIntegrationItem_domainName :: Lens' ListIntegrationItem Text

The unique name of the domain.

listIntegrationItem_uri :: Lens' ListIntegrationItem Text

The URI of the S3 bucket or any other type of data source.

listIntegrationItem_objectTypeName :: Lens' ListIntegrationItem Text

The name of the profile object type.

listIntegrationItem_createdAt :: Lens' ListIntegrationItem UTCTime

The timestamp of when the domain was created.

listIntegrationItem_lastUpdatedAt :: Lens' ListIntegrationItem UTCTime

The timestamp of when the domain was most recently edited.

ListProfileObjectTypeItem

listProfileObjectTypeItem_lastUpdatedAt :: Lens' ListProfileObjectTypeItem (Maybe UTCTime)

The timestamp of when the domain was most recently edited.

listProfileObjectTypeItem_createdAt :: Lens' ListProfileObjectTypeItem (Maybe UTCTime)

The timestamp of when the domain was created.

listProfileObjectTypeItem_tags :: Lens' ListProfileObjectTypeItem (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

listProfileObjectTypeItem_objectTypeName :: Lens' ListProfileObjectTypeItem Text

The name of the profile object type.

listProfileObjectTypeItem_description :: Lens' ListProfileObjectTypeItem Text

Description of the profile object type.

ListProfileObjectTypeTemplateItem

listProfileObjectTypeTemplateItem_sourceName :: Lens' ListProfileObjectTypeTemplateItem (Maybe Text)

The name of the source of the object template.

listProfileObjectTypeTemplateItem_templateId :: Lens' ListProfileObjectTypeTemplateItem (Maybe Text)

A unique identifier for the object template.

listProfileObjectTypeTemplateItem_sourceObject :: Lens' ListProfileObjectTypeTemplateItem (Maybe Text)

The source of the object template.

ListProfileObjectsItem

listProfileObjectsItem_profileObjectUniqueKey :: Lens' ListProfileObjectsItem (Maybe Text)

The unique identifier of the ProfileObject generated by the service.

listProfileObjectsItem_objectTypeName :: Lens' ListProfileObjectsItem (Maybe Text)

Specifies the kind of object being added to a profile, such as "Salesforce-Account."

listProfileObjectsItem_object :: Lens' ListProfileObjectsItem (Maybe Text)

A JSON representation of a ProfileObject that belongs to a profile.

MarketoSourceProperties

marketoSourceProperties_object :: Lens' MarketoSourceProperties Text

The object specified in the Marketo flow source.

MatchItem

matchItem_profileIds :: Lens' MatchItem (Maybe [Text])

A list of identifiers for profiles that match.

matchItem_matchId :: Lens' MatchItem (Maybe Text)

The unique identifiers for this group of profiles that match.

MatchingRequest

matchingRequest_enabled :: Lens' MatchingRequest Bool

The flag that enables the matching process of duplicate profiles.

MatchingResponse

matchingResponse_enabled :: Lens' MatchingResponse (Maybe Bool)

The flag that enables the matching process of duplicate profiles.

ObjectFilter

objectFilter_keyName :: Lens' ObjectFilter Text

A searchable identifier of a standard profile object. The predefined keys you can use to search for _asset include: _assetId, _assetName, _serialNumber. The predefined keys you can use to search for _case include: _caseId.

objectFilter_values :: Lens' ObjectFilter [Text]

A list of key values.

ObjectTypeField

objectTypeField_source :: Lens' ObjectTypeField (Maybe Text)

A field of a ProfileObject. For example: _source.FirstName, where “_source” is a ProfileObjectType of a Zendesk user and “FirstName” is a field in that ObjectType.

objectTypeField_contentType :: Lens' ObjectTypeField (Maybe FieldContentType)

The content type of the field. Used for determining equality when searching.

objectTypeField_target :: Lens' ObjectTypeField (Maybe Text)

The location of the data in the standard ProfileObject model. For example: _profile.Address.PostalCode.

ObjectTypeKey

objectTypeKey_fieldNames :: Lens' ObjectTypeKey (Maybe [Text])

The reference for the key name of the fields map.

objectTypeKey_standardIdentifiers :: Lens' ObjectTypeKey (Maybe [StandardIdentifier])

The types of keys that a ProfileObject can have. Each ProfileObject can have only 1 UNIQUE key but multiple PROFILE keys. PROFILE, ASSET or CASE means that this key can be used to tie an object to a PROFILE, ASSET or CASE respectively. UNIQUE means that it can be used to uniquely identify an object. If a key a is marked as SECONDARY, it will be used to search for profiles after all other PROFILE keys have been searched. A LOOKUP_ONLY key is only used to match a profile but is not persisted to be used for searching of the profile. A NEW_ONLY key is only used if the profile does not already exist before the object is ingested, otherwise it is only used for matching objects to profiles.

Profile

profile_shippingAddress :: Lens' Profile (Maybe Address)

The customer’s shipping address.

profile_mobilePhoneNumber :: Lens' Profile (Maybe Text)

The customer’s mobile phone number.

profile_mailingAddress :: Lens' Profile (Maybe Address)

The customer’s mailing address.

profile_middleName :: Lens' Profile (Maybe Text)

The customer’s middle name.

profile_personalEmailAddress :: Lens' Profile (Maybe Text)

The customer’s personal email address.

profile_lastName :: Lens' Profile (Maybe Text)

The customer’s last name.

profile_additionalInformation :: Lens' Profile (Maybe Text)

Any additional information relevant to the customer’s profile.

profile_homePhoneNumber :: Lens' Profile (Maybe Text)

The customer’s home phone number.

profile_address :: Lens' Profile (Maybe Address)

A generic address associated with the customer that is not mailing, shipping, or billing.

profile_partyType :: Lens' Profile (Maybe PartyType)

The type of profile used to describe the customer.

profile_profileId :: Lens' Profile (Maybe Text)

The unique identifier of a customer profile.

profile_businessEmailAddress :: Lens' Profile (Maybe Text)

The customer’s business email address.

profile_attributes :: Lens' Profile (Maybe (HashMap Text Text))

A key value pair of attributes of a customer profile.

profile_gender :: Lens' Profile (Maybe Gender)

The gender with which the customer identifies.

profile_phoneNumber :: Lens' Profile (Maybe Text)

The customer's phone number, which has not been specified as a mobile, home, or business number.

profile_accountNumber :: Lens' Profile (Maybe Text)

A unique account number that you have given to the customer.

profile_emailAddress :: Lens' Profile (Maybe Text)

The customer’s email address, which has not been specified as a personal or business address.

profile_firstName :: Lens' Profile (Maybe Text)

The customer’s first name.

profile_billingAddress :: Lens' Profile (Maybe Address)

The customer’s billing address.

profile_businessPhoneNumber :: Lens' Profile (Maybe Text)

The customer’s home phone number.

profile_birthDate :: Lens' Profile (Maybe Text)

The customer’s birth date.

profile_businessName :: Lens' Profile (Maybe Text)

The name of the customer’s business.

S3SourceProperties

s3SourceProperties_bucketPrefix :: Lens' S3SourceProperties (Maybe Text)

The object key for the Amazon S3 bucket in which the source files are stored.

s3SourceProperties_bucketName :: Lens' S3SourceProperties Text

The Amazon S3 bucket name where the source files are stored.

SalesforceSourceProperties

salesforceSourceProperties_enableDynamicFieldUpdate :: Lens' SalesforceSourceProperties (Maybe Bool)

The flag that enables dynamic fetching of new (recently added) fields in the Salesforce objects while running a flow.

salesforceSourceProperties_includeDeletedRecords :: Lens' SalesforceSourceProperties (Maybe Bool)

Indicates whether Amazon AppFlow includes deleted files in the flow run.

salesforceSourceProperties_object :: Lens' SalesforceSourceProperties Text

The object specified in the Salesforce flow source.

ScheduledTriggerProperties

scheduledTriggerProperties_scheduleEndTime :: Lens' ScheduledTriggerProperties (Maybe UTCTime)

Specifies the scheduled end time for a scheduled-trigger flow.

scheduledTriggerProperties_scheduleOffset :: Lens' ScheduledTriggerProperties (Maybe Natural)

Specifies the optional offset that is added to the time interval for a schedule-triggered flow.

scheduledTriggerProperties_dataPullMode :: Lens' ScheduledTriggerProperties (Maybe DataPullMode)

Specifies whether a scheduled flow has an incremental data transfer or a complete data transfer for each flow run.

scheduledTriggerProperties_scheduleStartTime :: Lens' ScheduledTriggerProperties (Maybe UTCTime)

Specifies the scheduled start time for a scheduled-trigger flow.

scheduledTriggerProperties_timezone :: Lens' ScheduledTriggerProperties (Maybe Text)

Specifies the time zone used when referring to the date and time of a scheduled-triggered flow, such as America/New_York.

scheduledTriggerProperties_firstExecutionFrom :: Lens' ScheduledTriggerProperties (Maybe UTCTime)

Specifies the date range for the records to import from the connector in the first flow run.

scheduledTriggerProperties_scheduleExpression :: Lens' ScheduledTriggerProperties Text

The scheduling expression that determines the rate at which the schedule will run, for example rate (5 minutes).

ServiceNowSourceProperties

serviceNowSourceProperties_object :: Lens' ServiceNowSourceProperties Text

The object specified in the ServiceNow flow source.

SourceConnectorProperties

sourceConnectorProperties_serviceNow :: Lens' SourceConnectorProperties (Maybe ServiceNowSourceProperties)

The properties that are applied when ServiceNow is being used as a source.

sourceConnectorProperties_marketo :: Lens' SourceConnectorProperties (Maybe MarketoSourceProperties)

The properties that are applied when Marketo is being used as a source.

sourceConnectorProperties_salesforce :: Lens' SourceConnectorProperties (Maybe SalesforceSourceProperties)

The properties that are applied when Salesforce is being used as a source.

sourceConnectorProperties_zendesk :: Lens' SourceConnectorProperties (Maybe ZendeskSourceProperties)

The properties that are applied when using Zendesk as a flow source.

sourceConnectorProperties_s3 :: Lens' SourceConnectorProperties (Maybe S3SourceProperties)

The properties that are applied when Amazon S3 is being used as the flow source.

SourceFlowConfig

sourceFlowConfig_connectorProfileName :: Lens' SourceFlowConfig (Maybe Text)

The name of the AppFlow connector profile. This name must be unique for each connector profile in the AWS account.

sourceFlowConfig_incrementalPullConfig :: Lens' SourceFlowConfig (Maybe IncrementalPullConfig)

Defines the configuration for a scheduled incremental data pull. If a valid configuration is provided, the fields specified in the configuration are used when querying for the incremental data pull.

sourceFlowConfig_connectorType :: Lens' SourceFlowConfig SourceConnectorType

The type of connector, such as Salesforce, Marketo, and so on.

sourceFlowConfig_sourceConnectorProperties :: Lens' SourceFlowConfig SourceConnectorProperties

Specifies the information that is required to query a particular source connector.

Task

task_taskProperties :: Lens' Task (Maybe (HashMap OperatorPropertiesKeys Text))

A map used to store task-related information. The service looks for particular information based on the TaskType.

task_connectorOperator :: Lens' Task (Maybe ConnectorOperator)

The operation to be performed on the provided source fields.

task_destinationField :: Lens' Task (Maybe Text)

A field in a destination connector, or a field value against which Amazon AppFlow validates a source field.

task_sourceFields :: Lens' Task [Text]

The source fields to which a particular task is applied.

task_taskType :: Lens' Task TaskType

Specifies the particular task implementation that Amazon AppFlow performs.

TriggerConfig

triggerConfig_triggerProperties :: Lens' TriggerConfig (Maybe TriggerProperties)

Specifies the configuration details of a schedule-triggered flow that you define. Currently, these settings only apply to the Scheduled trigger type.

triggerConfig_triggerType :: Lens' TriggerConfig TriggerType

Specifies the type of flow trigger. It can be OnDemand, Scheduled, or Event.

TriggerProperties

triggerProperties_scheduled :: Lens' TriggerProperties (Maybe ScheduledTriggerProperties)

Specifies the configuration details of a schedule-triggered flow that you define.

UpdateAddress

updateAddress_address2 :: Lens' UpdateAddress (Maybe Text)

The second line of a customer address.

updateAddress_state :: Lens' UpdateAddress (Maybe Text)

The state in which a customer lives.

updateAddress_address4 :: Lens' UpdateAddress (Maybe Text)

The fourth line of a customer address.

updateAddress_county :: Lens' UpdateAddress (Maybe Text)

The county in which a customer lives.

updateAddress_address3 :: Lens' UpdateAddress (Maybe Text)

The third line of a customer address.

updateAddress_postalCode :: Lens' UpdateAddress (Maybe Text)

The postal code of a customer address.

updateAddress_country :: Lens' UpdateAddress (Maybe Text)

The country in which a customer lives.

updateAddress_city :: Lens' UpdateAddress (Maybe Text)

The city in which a customer lives.

updateAddress_address1 :: Lens' UpdateAddress (Maybe Text)

The first line of a customer address.

updateAddress_province :: Lens' UpdateAddress (Maybe Text)

The province in which a customer lives.

ZendeskSourceProperties

zendeskSourceProperties_object :: Lens' ZendeskSourceProperties Text

The object specified in the Zendesk flow source.