Network.AWS.CustomerProfiles.ListProfileObjects

Description

Returns a list of objects associated with a profile of a given ProfileObjectType.

Creating a Request

data ListProfileObjects

See: newListProfileObjects smart constructor.

Constructors

ListProfileObjects'

Fields

objectFilter :: Maybe ObjectFilter

Applies a filter to the response to include profile objects with the specified index values. This filter is only supported for ObjectTypeName _asset and _case.

nextToken :: Maybe Text

The pagination token from the previous call to ListProfileObjects.

maxResults :: Maybe Natural

The maximum number of objects returned per page.

domainName :: Text

The unique name of the domain.

objectTypeName :: Text

The name of the profile object type.

profileId :: Text

The unique identifier of a customer profile.

newListProfileObjects

Arguments

:: Text
$sel:domainName:ListProfileObjects'

-> Text
$sel:objectTypeName:ListProfileObjects'

-> Text
$sel:profileId:ListProfileObjects'

-> ListProfileObjects

Create a value of ListProfileObjects with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:objectFilter:ListProfileObjects', listProfileObjects_objectFilter - Applies a filter to the response to include profile objects with the specified index values. This filter is only supported for ObjectTypeName _asset and _case.

$sel:nextToken:ListProfileObjects', listProfileObjects_nextToken - The pagination token from the previous call to ListProfileObjects.

$sel:maxResults:ListProfileObjects', listProfileObjects_maxResults - The maximum number of objects returned per page.

$sel:domainName:ListProfileObjects', listProfileObjects_domainName - The unique name of the domain.

$sel:objectTypeName:ListProfileObjects', listProfileObjects_objectTypeName - The name of the profile object type.

$sel:profileId:ListProfileObjects', listProfileObjects_profileId - The unique identifier of a customer profile.

Request Lenses

listProfileObjects_objectFilter :: Lens' ListProfileObjects (Maybe ObjectFilter)

Applies a filter to the response to include profile objects with the specified index values. This filter is only supported for ObjectTypeName _asset and _case.

listProfileObjects_nextToken :: Lens' ListProfileObjects (Maybe Text)

The pagination token from the previous call to ListProfileObjects.

listProfileObjects_maxResults :: Lens' ListProfileObjects (Maybe Natural)

The maximum number of objects returned per page.

listProfileObjects_domainName :: Lens' ListProfileObjects Text

The unique name of the domain.

listProfileObjects_objectTypeName :: Lens' ListProfileObjects Text

The name of the profile object type.

listProfileObjects_profileId :: Lens' ListProfileObjects Text

The unique identifier of a customer profile.

Destructuring the Response

data ListProfileObjectsResponse

See: newListProfileObjectsResponse smart constructor.

Constructors

ListProfileObjectsResponse'

Fields

items :: Maybe [ListProfileObjectsItem]

The list of ListProfileObject instances.

nextToken :: Maybe Text

The pagination token from the previous call to ListProfileObjects.

httpStatus :: Int

The response's http status code.

newListProfileObjectsResponse

Arguments

:: Int
$sel:httpStatus:ListProfileObjectsResponse'

-> ListProfileObjectsResponse

Create a value of ListProfileObjectsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:items:ListProfileObjectsResponse', listProfileObjectsResponse_items - The list of ListProfileObject instances.

$sel:nextToken:ListProfileObjects', listProfileObjectsResponse_nextToken - The pagination token from the previous call to ListProfileObjects.

$sel:httpStatus:ListProfileObjectsResponse', listProfileObjectsResponse_httpStatus - The response's http status code.

Response Lenses

listProfileObjectsResponse_items :: Lens' ListProfileObjectsResponse (Maybe [ListProfileObjectsItem])

The list of ListProfileObject instances.

listProfileObjectsResponse_nextToken :: Lens' ListProfileObjectsResponse (Maybe Text)

The pagination token from the previous call to ListProfileObjects.

listProfileObjectsResponse_httpStatus :: Lens' ListProfileObjectsResponse Int

The response's http status code.