Network.AWS.CustomerProfiles.PutProfileObject

Description

Adds additional objects to customer profiles of a given ObjectType.

When adding a specific profile object, like a Contact Trace Record (CTR), an inferred profile can get created if it is not mapped to an existing profile. The resulting profile will only have a phone number populated in the standard ProfileObject. Any additional CTRs with the same phone number will be mapped to the same inferred profile.

When a ProfileObject is created and if a ProfileObjectType already exists for the ProfileObject, it will provide data to a standard profile depending on the ProfileObjectType definition.

PutProfileObject needs an ObjectType, which can be created using PutProfileObjectType.

Creating a Request

data PutProfileObject

See: newPutProfileObject smart constructor.

Constructors

PutProfileObject'

Fields

objectTypeName :: Text

The name of the profile object type.

object' :: Text

A string that is serialized from a JSON object.

domainName :: Text

The unique name of the domain.

newPutProfileObject

Arguments

:: Text
$sel:objectTypeName:PutProfileObject'

-> Text
$sel:object':PutProfileObject'

-> Text
$sel:domainName:PutProfileObject'

-> PutProfileObject

Create a value of PutProfileObject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:objectTypeName:PutProfileObject', putProfileObject_objectTypeName - The name of the profile object type.

$sel:object':PutProfileObject', putProfileObject_object - A string that is serialized from a JSON object.

$sel:domainName:PutProfileObject', putProfileObject_domainName - The unique name of the domain.

Request Lenses

putProfileObject_objectTypeName :: Lens' PutProfileObject Text

The name of the profile object type.

putProfileObject_object :: Lens' PutProfileObject Text

A string that is serialized from a JSON object.

putProfileObject_domainName :: Lens' PutProfileObject Text

The unique name of the domain.

Destructuring the Response

data PutProfileObjectResponse

See: newPutProfileObjectResponse smart constructor.

Constructors

PutProfileObjectResponse'

Fields

profileObjectUniqueKey :: Maybe Text

The unique identifier of the profile object generated by the service.

httpStatus :: Int

The response's http status code.

newPutProfileObjectResponse

Arguments

:: Int
$sel:httpStatus:PutProfileObjectResponse'

-> PutProfileObjectResponse

Create a value of PutProfileObjectResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:profileObjectUniqueKey:PutProfileObjectResponse', putProfileObjectResponse_profileObjectUniqueKey - The unique identifier of the profile object generated by the service.

$sel:httpStatus:PutProfileObjectResponse', putProfileObjectResponse_httpStatus - The response's http status code.

Response Lenses

putProfileObjectResponse_profileObjectUniqueKey :: Lens' PutProfileObjectResponse (Maybe Text)

The unique identifier of the profile object generated by the service.

putProfileObjectResponse_httpStatus :: Lens' PutProfileObjectResponse Int

The response's http status code.