Network.AWS.CustomerProfiles.Types.ConnectorOperator

Description

Documentation

data ConnectorOperator

The operation to be performed on the provided source fields.

See: newConnectorOperator smart constructor.

Constructors

ConnectorOperator'

Fields

serviceNow :: Maybe ServiceNowConnectorOperator

The operation to be performed on the provided ServiceNow source fields.

marketo :: Maybe MarketoConnectorOperator

The operation to be performed on the provided Marketo source fields.

salesforce :: Maybe SalesforceConnectorOperator

The operation to be performed on the provided Salesforce source fields.

zendesk :: Maybe ZendeskConnectorOperator

The operation to be performed on the provided Zendesk source fields.

s3 :: Maybe S3ConnectorOperator

The operation to be performed on the provided Amazon S3 source fields.

newConnectorOperator :: ConnectorOperator

Create a value of ConnectorOperator with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceNow:ConnectorOperator', connectorOperator_serviceNow - The operation to be performed on the provided ServiceNow source fields.

$sel:marketo:ConnectorOperator', connectorOperator_marketo - The operation to be performed on the provided Marketo source fields.

$sel:salesforce:ConnectorOperator', connectorOperator_salesforce - The operation to be performed on the provided Salesforce source fields.

$sel:zendesk:ConnectorOperator', connectorOperator_zendesk - The operation to be performed on the provided Zendesk source fields.

$sel:s3:ConnectorOperator', connectorOperator_s3 - The operation to be performed on the provided Amazon S3 source fields.

connectorOperator_serviceNow :: Lens' ConnectorOperator (Maybe ServiceNowConnectorOperator)

The operation to be performed on the provided ServiceNow source fields.

connectorOperator_marketo :: Lens' ConnectorOperator (Maybe MarketoConnectorOperator)

The operation to be performed on the provided Marketo source fields.

connectorOperator_salesforce :: Lens' ConnectorOperator (Maybe SalesforceConnectorOperator)

The operation to be performed on the provided Salesforce source fields.

connectorOperator_zendesk :: Lens' ConnectorOperator (Maybe ZendeskConnectorOperator)

The operation to be performed on the provided Zendesk source fields.

connectorOperator_s3 :: Lens' ConnectorOperator (Maybe S3ConnectorOperator)

The operation to be performed on the provided Amazon S3 source fields.