Network.AWS.CustomerProfiles.Types.DomainStats

Description

Documentation

data DomainStats

Usage-specific statistics about the domain.

See: newDomainStats smart constructor.

Constructors

DomainStats'

Fields

meteringProfileCount :: Maybe Integer

The number of profiles that you are currently paying for in the domain. If you have more than 100 objects associated with a single profile, that profile counts as two profiles. If you have more than 200 objects, that profile counts as three, and so on.

totalSize :: Maybe Integer

The total size, in bytes, of all objects in the domain.

profileCount :: Maybe Integer

The total number of profiles currently in the domain.

objectCount :: Maybe Integer

The total number of objects in domain.

newDomainStats :: DomainStats

Create a value of DomainStats with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:meteringProfileCount:DomainStats', domainStats_meteringProfileCount - The number of profiles that you are currently paying for in the domain. If you have more than 100 objects associated with a single profile, that profile counts as two profiles. If you have more than 200 objects, that profile counts as three, and so on.

$sel:totalSize:DomainStats', domainStats_totalSize - The total size, in bytes, of all objects in the domain.

$sel:profileCount:DomainStats', domainStats_profileCount - The total number of profiles currently in the domain.

$sel:objectCount:DomainStats', domainStats_objectCount - The total number of objects in domain.

domainStats_meteringProfileCount :: Lens' DomainStats (Maybe Integer)

The number of profiles that you are currently paying for in the domain. If you have more than 100 objects associated with a single profile, that profile counts as two profiles. If you have more than 200 objects, that profile counts as three, and so on.

domainStats_totalSize :: Lens' DomainStats (Maybe Integer)

The total size, in bytes, of all objects in the domain.

domainStats_profileCount :: Lens' DomainStats (Maybe Integer)

The total number of profiles currently in the domain.

domainStats_objectCount :: Lens' DomainStats (Maybe Integer)

The total number of objects in domain.