Network.AWS.CustomerProfiles.Types.ListIntegrationItem

Description

Documentation

data ListIntegrationItem

An integration in list of integrations.

See: newListIntegrationItem smart constructor.

Constructors

ListIntegrationItem'

Fields

tags :: Maybe (HashMap Text Text)

The tags used to organize, track, or control access for this resource.

domainName :: Text

The unique name of the domain.

uri :: Text

The URI of the S3 bucket or any other type of data source.

objectTypeName :: Text

The name of the profile object type.

createdAt :: POSIX

The timestamp of when the domain was created.

lastUpdatedAt :: POSIX

The timestamp of when the domain was most recently edited.

newListIntegrationItem

Arguments

:: Text
$sel:domainName:ListIntegrationItem'

-> Text
$sel:uri:ListIntegrationItem'

-> Text
$sel:objectTypeName:ListIntegrationItem'

-> UTCTime
$sel:createdAt:ListIntegrationItem'

-> UTCTime
$sel:lastUpdatedAt:ListIntegrationItem'

-> ListIntegrationItem

Create a value of ListIntegrationItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:ListIntegrationItem', listIntegrationItem_tags - The tags used to organize, track, or control access for this resource.

$sel:domainName:ListIntegrationItem', listIntegrationItem_domainName - The unique name of the domain.

$sel:uri:ListIntegrationItem', listIntegrationItem_uri - The URI of the S3 bucket or any other type of data source.

$sel:objectTypeName:ListIntegrationItem', listIntegrationItem_objectTypeName - The name of the profile object type.

$sel:createdAt:ListIntegrationItem', listIntegrationItem_createdAt - The timestamp of when the domain was created.

$sel:lastUpdatedAt:ListIntegrationItem', listIntegrationItem_lastUpdatedAt - The timestamp of when the domain was most recently edited.

listIntegrationItem_tags :: Lens' ListIntegrationItem (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

listIntegrationItem_domainName :: Lens' ListIntegrationItem Text

The unique name of the domain.

listIntegrationItem_uri :: Lens' ListIntegrationItem Text

The URI of the S3 bucket or any other type of data source.

listIntegrationItem_objectTypeName :: Lens' ListIntegrationItem Text

The name of the profile object type.

listIntegrationItem_createdAt :: Lens' ListIntegrationItem UTCTime

The timestamp of when the domain was created.

listIntegrationItem_lastUpdatedAt :: Lens' ListIntegrationItem UTCTime

The timestamp of when the domain was most recently edited.