Network.AWS.CustomerProfiles.Types.MarketoSourceProperties

Description

Documentation

data MarketoSourceProperties

The properties that are applied when Marketo is being used as a source.

See: newMarketoSourceProperties smart constructor.

Constructors

MarketoSourceProperties'

Fields

object' :: Text

The object specified in the Marketo flow source.

newMarketoSourceProperties

Arguments

:: Text
$sel:object':MarketoSourceProperties'

-> MarketoSourceProperties

Create a value of MarketoSourceProperties with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:object':MarketoSourceProperties', marketoSourceProperties_object - The object specified in the Marketo flow source.

marketoSourceProperties_object :: Lens' MarketoSourceProperties Text

The object specified in the Marketo flow source.