Network.AWS.CustomerProfiles.Types.ObjectTypeField

Description

Documentation

data ObjectTypeField

Represents a field in a ProfileObjectType.

See: newObjectTypeField smart constructor.

Constructors

ObjectTypeField'

Fields

source :: Maybe Text

A field of a ProfileObject. For example: _source.FirstName, where “_source” is a ProfileObjectType of a Zendesk user and “FirstName” is a field in that ObjectType.

contentType :: Maybe FieldContentType

The content type of the field. Used for determining equality when searching.

target :: Maybe Text

The location of the data in the standard ProfileObject model. For example: _profile.Address.PostalCode.

newObjectTypeField :: ObjectTypeField

Create a value of ObjectTypeField with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:source:ObjectTypeField', objectTypeField_source - A field of a ProfileObject. For example: _source.FirstName, where “_source” is a ProfileObjectType of a Zendesk user and “FirstName” is a field in that ObjectType.

$sel:contentType:ObjectTypeField', objectTypeField_contentType - The content type of the field. Used for determining equality when searching.

$sel:target:ObjectTypeField', objectTypeField_target - The location of the data in the standard ProfileObject model. For example: _profile.Address.PostalCode.

objectTypeField_source :: Lens' ObjectTypeField (Maybe Text)

A field of a ProfileObject. For example: _source.FirstName, where “_source” is a ProfileObjectType of a Zendesk user and “FirstName” is a field in that ObjectType.

objectTypeField_contentType :: Lens' ObjectTypeField (Maybe FieldContentType)

The content type of the field. Used for determining equality when searching.

objectTypeField_target :: Lens' ObjectTypeField (Maybe Text)

The location of the data in the standard ProfileObject model. For example: _profile.Address.PostalCode.