Network.AWS.CustomerProfiles.Types.SourceConnectorProperties

Description

Documentation

data SourceConnectorProperties

Specifies the information that is required to query a particular Amazon AppFlow connector. Customer Profiles supports Salesforce, Zendesk, Marketo, ServiceNow and Amazon S3.

See: newSourceConnectorProperties smart constructor.

Constructors

SourceConnectorProperties'

Fields

serviceNow :: Maybe ServiceNowSourceProperties

The properties that are applied when ServiceNow is being used as a source.

marketo :: Maybe MarketoSourceProperties

The properties that are applied when Marketo is being used as a source.

salesforce :: Maybe SalesforceSourceProperties

The properties that are applied when Salesforce is being used as a source.

zendesk :: Maybe ZendeskSourceProperties

The properties that are applied when using Zendesk as a flow source.

s3 :: Maybe S3SourceProperties

The properties that are applied when Amazon S3 is being used as the flow source.

newSourceConnectorProperties :: SourceConnectorProperties

Create a value of SourceConnectorProperties with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceNow:SourceConnectorProperties', sourceConnectorProperties_serviceNow - The properties that are applied when ServiceNow is being used as a source.

$sel:marketo:SourceConnectorProperties', sourceConnectorProperties_marketo - The properties that are applied when Marketo is being used as a source.

$sel:salesforce:SourceConnectorProperties', sourceConnectorProperties_salesforce - The properties that are applied when Salesforce is being used as a source.

$sel:zendesk:SourceConnectorProperties', sourceConnectorProperties_zendesk - The properties that are applied when using Zendesk as a flow source.

$sel:s3:SourceConnectorProperties', sourceConnectorProperties_s3 - The properties that are applied when Amazon S3 is being used as the flow source.

sourceConnectorProperties_serviceNow :: Lens' SourceConnectorProperties (Maybe ServiceNowSourceProperties)

The properties that are applied when ServiceNow is being used as a source.

sourceConnectorProperties_marketo :: Lens' SourceConnectorProperties (Maybe MarketoSourceProperties)

The properties that are applied when Marketo is being used as a source.

sourceConnectorProperties_salesforce :: Lens' SourceConnectorProperties (Maybe SalesforceSourceProperties)

The properties that are applied when Salesforce is being used as a source.

sourceConnectorProperties_zendesk :: Lens' SourceConnectorProperties (Maybe ZendeskSourceProperties)

The properties that are applied when using Zendesk as a flow source.

sourceConnectorProperties_s3 :: Lens' SourceConnectorProperties (Maybe S3SourceProperties)

The properties that are applied when Amazon S3 is being used as the flow source.