Network.AWS.CustomerProfiles.Types.SourceFlowConfig

Description

Documentation

data SourceFlowConfig

Contains information about the configuration of the source connector used in the flow.

See: newSourceFlowConfig smart constructor.

Constructors

SourceFlowConfig'

Fields

connectorProfileName :: Maybe Text

The name of the AppFlow connector profile. This name must be unique for each connector profile in the AWS account.

incrementalPullConfig :: Maybe IncrementalPullConfig

Defines the configuration for a scheduled incremental data pull. If a valid configuration is provided, the fields specified in the configuration are used when querying for the incremental data pull.

connectorType :: SourceConnectorType

The type of connector, such as Salesforce, Marketo, and so on.

sourceConnectorProperties :: SourceConnectorProperties

Specifies the information that is required to query a particular source connector.

newSourceFlowConfig

Arguments

:: SourceConnectorType
$sel:connectorType:SourceFlowConfig'

-> SourceConnectorProperties
$sel:sourceConnectorProperties:SourceFlowConfig'

-> SourceFlowConfig

Create a value of SourceFlowConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:connectorProfileName:SourceFlowConfig', sourceFlowConfig_connectorProfileName - The name of the AppFlow connector profile. This name must be unique for each connector profile in the AWS account.

$sel:incrementalPullConfig:SourceFlowConfig', sourceFlowConfig_incrementalPullConfig - Defines the configuration for a scheduled incremental data pull. If a valid configuration is provided, the fields specified in the configuration are used when querying for the incremental data pull.

$sel:connectorType:SourceFlowConfig', sourceFlowConfig_connectorType - The type of connector, such as Salesforce, Marketo, and so on.

$sel:sourceConnectorProperties:SourceFlowConfig', sourceFlowConfig_sourceConnectorProperties - Specifies the information that is required to query a particular source connector.

sourceFlowConfig_connectorProfileName :: Lens' SourceFlowConfig (Maybe Text)

The name of the AppFlow connector profile. This name must be unique for each connector profile in the AWS account.

sourceFlowConfig_incrementalPullConfig :: Lens' SourceFlowConfig (Maybe IncrementalPullConfig)

Defines the configuration for a scheduled incremental data pull. If a valid configuration is provided, the fields specified in the configuration are used when querying for the incremental data pull.

sourceFlowConfig_connectorType :: Lens' SourceFlowConfig SourceConnectorType

The type of connector, such as Salesforce, Marketo, and so on.

sourceFlowConfig_sourceConnectorProperties :: Lens' SourceFlowConfig SourceConnectorProperties

Specifies the information that is required to query a particular source connector.