Network.AWS.CustomerProfiles.UpdateProfile

Description

Updates the properties of a profile. The ProfileId is required for updating a customer profile.

When calling the UpdateProfile API, specifying an empty string value means that any existing value will be removed. Not specifying a string value means that any value already there will be kept.

Creating a Request

data UpdateProfile

See: newUpdateProfile smart constructor.

Constructors

UpdateProfile'

Fields

shippingAddress :: Maybe UpdateAddress

The customer’s shipping address.

mobilePhoneNumber :: Maybe Text

The customer’s mobile phone number.

mailingAddress :: Maybe UpdateAddress

The customer’s mailing address.

middleName :: Maybe Text

The customer’s middle name.

personalEmailAddress :: Maybe Text

The customer’s personal email address.

lastName :: Maybe Text

The customer’s last name.

additionalInformation :: Maybe Text

Any additional information relevant to the customer’s profile.

homePhoneNumber :: Maybe Text

The customer’s home phone number.

address :: Maybe UpdateAddress

A generic address associated with the customer that is not mailing, shipping, or billing.

partyType :: Maybe PartyType

The type of profile used to describe the customer.

businessEmailAddress :: Maybe Text

The customer’s business email address.

attributes :: Maybe (HashMap Text Text)

A key value pair of attributes of a customer profile.

gender :: Maybe Gender

The gender with which the customer identifies.

phoneNumber :: Maybe Text

The customer’s phone number, which has not been specified as a mobile, home, or business number.

accountNumber :: Maybe Text

A unique account number that you have given to the customer.

emailAddress :: Maybe Text

The customer’s email address, which has not been specified as a personal or business address.

firstName :: Maybe Text

The customer’s first name.

billingAddress :: Maybe UpdateAddress

The customer’s billing address.

businessPhoneNumber :: Maybe Text

The customer’s business phone number.

birthDate :: Maybe Text

The customer’s birth date.

businessName :: Maybe Text

The name of the customer’s business.

domainName :: Text

The unique name of the domain.

profileId :: Text

The unique identifier of a customer profile.

newUpdateProfile

Arguments

:: Text
$sel:domainName:UpdateProfile'

-> Text
$sel:profileId:UpdateProfile'

-> UpdateProfile

Create a value of UpdateProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:shippingAddress:UpdateProfile', updateProfile_shippingAddress - The customer’s shipping address.

$sel:mobilePhoneNumber:UpdateProfile', updateProfile_mobilePhoneNumber - The customer’s mobile phone number.

$sel:mailingAddress:UpdateProfile', updateProfile_mailingAddress - The customer’s mailing address.

$sel:middleName:UpdateProfile', updateProfile_middleName - The customer’s middle name.

$sel:personalEmailAddress:UpdateProfile', updateProfile_personalEmailAddress - The customer’s personal email address.

$sel:lastName:UpdateProfile', updateProfile_lastName - The customer’s last name.

$sel:additionalInformation:UpdateProfile', updateProfile_additionalInformation - Any additional information relevant to the customer’s profile.

$sel:homePhoneNumber:UpdateProfile', updateProfile_homePhoneNumber - The customer’s home phone number.

$sel:address:UpdateProfile', updateProfile_address - A generic address associated with the customer that is not mailing, shipping, or billing.

$sel:partyType:UpdateProfile', updateProfile_partyType - The type of profile used to describe the customer.

$sel:businessEmailAddress:UpdateProfile', updateProfile_businessEmailAddress - The customer’s business email address.

$sel:attributes:UpdateProfile', updateProfile_attributes - A key value pair of attributes of a customer profile.

$sel:gender:UpdateProfile', updateProfile_gender - The gender with which the customer identifies.

$sel:phoneNumber:UpdateProfile', updateProfile_phoneNumber - The customer’s phone number, which has not been specified as a mobile, home, or business number.

$sel:accountNumber:UpdateProfile', updateProfile_accountNumber - A unique account number that you have given to the customer.

$sel:emailAddress:UpdateProfile', updateProfile_emailAddress - The customer’s email address, which has not been specified as a personal or business address.

$sel:firstName:UpdateProfile', updateProfile_firstName - The customer’s first name.

$sel:billingAddress:UpdateProfile', updateProfile_billingAddress - The customer’s billing address.

$sel:businessPhoneNumber:UpdateProfile', updateProfile_businessPhoneNumber - The customer’s business phone number.

$sel:birthDate:UpdateProfile', updateProfile_birthDate - The customer’s birth date.

$sel:businessName:UpdateProfile', updateProfile_businessName - The name of the customer’s business.

$sel:domainName:UpdateProfile', updateProfile_domainName - The unique name of the domain.

$sel:profileId:UpdateProfile', updateProfile_profileId - The unique identifier of a customer profile.

Request Lenses

updateProfile_shippingAddress :: Lens' UpdateProfile (Maybe UpdateAddress)

The customer’s shipping address.

updateProfile_mobilePhoneNumber :: Lens' UpdateProfile (Maybe Text)

The customer’s mobile phone number.

updateProfile_mailingAddress :: Lens' UpdateProfile (Maybe UpdateAddress)

The customer’s mailing address.

updateProfile_middleName :: Lens' UpdateProfile (Maybe Text)

The customer’s middle name.

updateProfile_personalEmailAddress :: Lens' UpdateProfile (Maybe Text)

The customer’s personal email address.

updateProfile_lastName :: Lens' UpdateProfile (Maybe Text)

The customer’s last name.

updateProfile_additionalInformation :: Lens' UpdateProfile (Maybe Text)

Any additional information relevant to the customer’s profile.

updateProfile_homePhoneNumber :: Lens' UpdateProfile (Maybe Text)

The customer’s home phone number.

updateProfile_address :: Lens' UpdateProfile (Maybe UpdateAddress)

A generic address associated with the customer that is not mailing, shipping, or billing.

updateProfile_partyType :: Lens' UpdateProfile (Maybe PartyType)

The type of profile used to describe the customer.

updateProfile_businessEmailAddress :: Lens' UpdateProfile (Maybe Text)

The customer’s business email address.

updateProfile_attributes :: Lens' UpdateProfile (Maybe (HashMap Text Text))

A key value pair of attributes of a customer profile.

updateProfile_gender :: Lens' UpdateProfile (Maybe Gender)

The gender with which the customer identifies.

updateProfile_phoneNumber :: Lens' UpdateProfile (Maybe Text)

The customer’s phone number, which has not been specified as a mobile, home, or business number.

updateProfile_accountNumber :: Lens' UpdateProfile (Maybe Text)

A unique account number that you have given to the customer.

updateProfile_emailAddress :: Lens' UpdateProfile (Maybe Text)

The customer’s email address, which has not been specified as a personal or business address.

updateProfile_firstName :: Lens' UpdateProfile (Maybe Text)

The customer’s first name.

updateProfile_billingAddress :: Lens' UpdateProfile (Maybe UpdateAddress)

The customer’s billing address.

updateProfile_businessPhoneNumber :: Lens' UpdateProfile (Maybe Text)

The customer’s business phone number.

updateProfile_birthDate :: Lens' UpdateProfile (Maybe Text)

The customer’s birth date.

updateProfile_businessName :: Lens' UpdateProfile (Maybe Text)

The name of the customer’s business.

updateProfile_domainName :: Lens' UpdateProfile Text

The unique name of the domain.

updateProfile_profileId :: Lens' UpdateProfile Text

The unique identifier of a customer profile.

Destructuring the Response

data UpdateProfileResponse

See: newUpdateProfileResponse smart constructor.

Constructors

UpdateProfileResponse'

Fields

httpStatus :: Int

The response's http status code.

profileId :: Text

The unique identifier of a customer profile.

newUpdateProfileResponse

Arguments

:: Int
$sel:httpStatus:UpdateProfileResponse'

-> Text
$sel:profileId:UpdateProfile'

-> UpdateProfileResponse

Create a value of UpdateProfileResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateProfileResponse', updateProfileResponse_httpStatus - The response's http status code.

$sel:profileId:UpdateProfile', updateProfileResponse_profileId - The unique identifier of a customer profile.

Response Lenses

updateProfileResponse_httpStatus :: Lens' UpdateProfileResponse Int

The response's http status code.

updateProfileResponse_profileId :: Lens' UpdateProfileResponse Text

The unique identifier of a customer profile.