Network.AWS.DAX.CreateCluster

Description

Creates a DAX cluster. All nodes in the cluster run the same DAX caching software.

Creating a Request

data CreateCluster

See: newCreateCluster smart constructor.

Constructors

CreateCluster'

Fields

securityGroupIds :: Maybe [Text]

A list of security group IDs to be assigned to each node in the DAX cluster. (Each of the security group ID is system-generated.)

If this parameter is not specified, DAX assigns the default VPC security group to each node.

sSESpecification :: Maybe SSESpecification

Represents the settings used to enable server-side encryption on the cluster.

subnetGroupName :: Maybe Text

The name of the subnet group to be used for the replication group.

DAX clusters can only run in an Amazon VPC environment. All of the subnets that you specify in a subnet group must exist in the same VPC.

clusterEndpointEncryptionType :: Maybe ClusterEndpointEncryptionType

The type of encryption the cluster's endpoint should support. Values are:

NONE for no encryption

TLS for Transport Layer Security

preferredMaintenanceWindow :: Maybe Text

Specifies the weekly time range during which maintenance on the DAX cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period. Valid values for ddd are:

sun

mon

tue

wed

thu

fri

sat

Example: sun:05:00-sun:09:00

If you don't specify a preferred maintenance window when you create or modify a cache cluster, DAX assigns a 60-minute maintenance window on a randomly selected day of the week.

availabilityZones :: Maybe [Text]

The Availability Zones (AZs) in which the cluster nodes will reside after the cluster has been created or updated. If provided, the length of this list must equal the ReplicationFactor parameter. If you omit this parameter, DAX will spread the nodes across Availability Zones for the highest availability.

description :: Maybe Text

A description of the cluster.

notificationTopicArn :: Maybe Text

The Amazon Resource Name (ARN) of the Amazon SNS topic to which notifications will be sent.

The Amazon SNS topic owner must be same as the DAX cluster owner.

tags :: Maybe [Tag]

A set of tags to associate with the DAX cluster.

parameterGroupName :: Maybe Text

The parameter group to be associated with the DAX cluster.

clusterName :: Text

The cluster identifier. This parameter is stored as a lowercase string.

Constraints:

A name must contain from 1 to 20 alphanumeric characters or hyphens.

The first character must be a letter.

A name cannot end with a hyphen or contain two consecutive hyphens.

nodeType :: Text

The compute and memory capacity of the nodes in the cluster.

replicationFactor :: Int

The number of nodes in the DAX cluster. A replication factor of 1 will create a single-node cluster, without any read replicas. For additional fault tolerance, you can create a multiple node cluster with one or more read replicas. To do this, set ReplicationFactor to a number between 3 (one primary and two read replicas) and 10 (one primary and nine read replicas). If the AvailabilityZones parameter is provided, its length must equal the ReplicationFactor.

AWS recommends that you have at least two read replicas per cluster.

iamRoleArn :: Text

A valid Amazon Resource Name (ARN) that identifies an IAM role. At runtime, DAX will assume this role and use the role's permissions to access DynamoDB on your behalf.

newCreateCluster

Arguments

:: Text
$sel:clusterName:CreateCluster'

-> Text
$sel:nodeType:CreateCluster'

-> Int
$sel:replicationFactor:CreateCluster'

-> Text
$sel:iamRoleArn:CreateCluster'

-> CreateCluster

Create a value of CreateCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:securityGroupIds:CreateCluster', createCluster_securityGroupIds - A list of security group IDs to be assigned to each node in the DAX cluster. (Each of the security group ID is system-generated.)

If this parameter is not specified, DAX assigns the default VPC security group to each node.

$sel:sSESpecification:CreateCluster', createCluster_sSESpecification - Represents the settings used to enable server-side encryption on the cluster.

$sel:subnetGroupName:CreateCluster', createCluster_subnetGroupName - The name of the subnet group to be used for the replication group.

DAX clusters can only run in an Amazon VPC environment. All of the subnets that you specify in a subnet group must exist in the same VPC.

$sel:clusterEndpointEncryptionType:CreateCluster', createCluster_clusterEndpointEncryptionType - The type of encryption the cluster's endpoint should support. Values are:

NONE for no encryption

TLS for Transport Layer Security

$sel:preferredMaintenanceWindow:CreateCluster', createCluster_preferredMaintenanceWindow - Specifies the weekly time range during which maintenance on the DAX cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period. Valid values for ddd are:

sun

mon

tue

wed

thu

fri

sat

Example: sun:05:00-sun:09:00

If you don't specify a preferred maintenance window when you create or modify a cache cluster, DAX assigns a 60-minute maintenance window on a randomly selected day of the week.

$sel:availabilityZones:CreateCluster', createCluster_availabilityZones - The Availability Zones (AZs) in which the cluster nodes will reside after the cluster has been created or updated. If provided, the length of this list must equal the ReplicationFactor parameter. If you omit this parameter, DAX will spread the nodes across Availability Zones for the highest availability.

$sel:description:CreateCluster', createCluster_description - A description of the cluster.

$sel:notificationTopicArn:CreateCluster', createCluster_notificationTopicArn - The Amazon Resource Name (ARN) of the Amazon SNS topic to which notifications will be sent.

The Amazon SNS topic owner must be same as the DAX cluster owner.

$sel:tags:CreateCluster', createCluster_tags - A set of tags to associate with the DAX cluster.

$sel:parameterGroupName:CreateCluster', createCluster_parameterGroupName - The parameter group to be associated with the DAX cluster.

$sel:clusterName:CreateCluster', createCluster_clusterName - The cluster identifier. This parameter is stored as a lowercase string.

Constraints:

A name must contain from 1 to 20 alphanumeric characters or hyphens.

The first character must be a letter.

A name cannot end with a hyphen or contain two consecutive hyphens.

$sel:nodeType:CreateCluster', createCluster_nodeType - The compute and memory capacity of the nodes in the cluster.

$sel:replicationFactor:CreateCluster', createCluster_replicationFactor - The number of nodes in the DAX cluster. A replication factor of 1 will create a single-node cluster, without any read replicas. For additional fault tolerance, you can create a multiple node cluster with one or more read replicas. To do this, set ReplicationFactor to a number between 3 (one primary and two read replicas) and 10 (one primary and nine read replicas). If the AvailabilityZones parameter is provided, its length must equal the ReplicationFactor.

AWS recommends that you have at least two read replicas per cluster.

$sel:iamRoleArn:CreateCluster', createCluster_iamRoleArn - A valid Amazon Resource Name (ARN) that identifies an IAM role. At runtime, DAX will assume this role and use the role's permissions to access DynamoDB on your behalf.

Request Lenses

createCluster_securityGroupIds :: Lens' CreateCluster (Maybe [Text])

A list of security group IDs to be assigned to each node in the DAX cluster. (Each of the security group ID is system-generated.)

If this parameter is not specified, DAX assigns the default VPC security group to each node.

createCluster_sSESpecification :: Lens' CreateCluster (Maybe SSESpecification)

Represents the settings used to enable server-side encryption on the cluster.

createCluster_subnetGroupName :: Lens' CreateCluster (Maybe Text)

The name of the subnet group to be used for the replication group.

DAX clusters can only run in an Amazon VPC environment. All of the subnets that you specify in a subnet group must exist in the same VPC.

createCluster_clusterEndpointEncryptionType :: Lens' CreateCluster (Maybe ClusterEndpointEncryptionType)

The type of encryption the cluster's endpoint should support. Values are:

NONE for no encryption

TLS for Transport Layer Security

createCluster_preferredMaintenanceWindow :: Lens' CreateCluster (Maybe Text)

Specifies the weekly time range during which maintenance on the DAX cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period. Valid values for ddd are:

sun

mon

tue

wed

thu

fri

sat

Example: sun:05:00-sun:09:00

If you don't specify a preferred maintenance window when you create or modify a cache cluster, DAX assigns a 60-minute maintenance window on a randomly selected day of the week.

createCluster_availabilityZones :: Lens' CreateCluster (Maybe [Text])

The Availability Zones (AZs) in which the cluster nodes will reside after the cluster has been created or updated. If provided, the length of this list must equal the ReplicationFactor parameter. If you omit this parameter, DAX will spread the nodes across Availability Zones for the highest availability.

createCluster_description :: Lens' CreateCluster (Maybe Text)

A description of the cluster.

createCluster_notificationTopicArn :: Lens' CreateCluster (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon SNS topic to which notifications will be sent.

The Amazon SNS topic owner must be same as the DAX cluster owner.

createCluster_tags :: Lens' CreateCluster (Maybe [Tag])

A set of tags to associate with the DAX cluster.

createCluster_parameterGroupName :: Lens' CreateCluster (Maybe Text)

The parameter group to be associated with the DAX cluster.

createCluster_clusterName :: Lens' CreateCluster Text

The cluster identifier. This parameter is stored as a lowercase string.

Constraints:

A name must contain from 1 to 20 alphanumeric characters or hyphens.

The first character must be a letter.

A name cannot end with a hyphen or contain two consecutive hyphens.

createCluster_nodeType :: Lens' CreateCluster Text

The compute and memory capacity of the nodes in the cluster.

createCluster_replicationFactor :: Lens' CreateCluster Int

The number of nodes in the DAX cluster. A replication factor of 1 will create a single-node cluster, without any read replicas. For additional fault tolerance, you can create a multiple node cluster with one or more read replicas. To do this, set ReplicationFactor to a number between 3 (one primary and two read replicas) and 10 (one primary and nine read replicas). If the AvailabilityZones parameter is provided, its length must equal the ReplicationFactor.

AWS recommends that you have at least two read replicas per cluster.

createCluster_iamRoleArn :: Lens' CreateCluster Text

A valid Amazon Resource Name (ARN) that identifies an IAM role. At runtime, DAX will assume this role and use the role's permissions to access DynamoDB on your behalf.

Destructuring the Response

data CreateClusterResponse

See: newCreateClusterResponse smart constructor.

Constructors

CreateClusterResponse'

Fields

cluster :: Maybe Cluster

A description of the DAX cluster that you have created.

httpStatus :: Int

The response's http status code.

newCreateClusterResponse

Arguments

:: Int
$sel:httpStatus:CreateClusterResponse'

-> CreateClusterResponse

Create a value of CreateClusterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cluster:CreateClusterResponse', createClusterResponse_cluster - A description of the DAX cluster that you have created.

$sel:httpStatus:CreateClusterResponse', createClusterResponse_httpStatus - The response's http status code.

Response Lenses

createClusterResponse_cluster :: Lens' CreateClusterResponse (Maybe Cluster)

A description of the DAX cluster that you have created.

createClusterResponse_httpStatus :: Lens' CreateClusterResponse Int

The response's http status code.