Network.AWS.DAX.CreateSubnetGroup

Description

Creates a new subnet group.

Creating a Request

data CreateSubnetGroup

See: newCreateSubnetGroup smart constructor.

Constructors

CreateSubnetGroup'

Fields

description :: Maybe Text

A description for the subnet group

subnetGroupName :: Text

A name for the subnet group. This value is stored as a lowercase string.

subnetIds :: [Text]

A list of VPC subnet IDs for the subnet group.

newCreateSubnetGroup

Arguments

:: Text
$sel:subnetGroupName:CreateSubnetGroup'

-> CreateSubnetGroup

Create a value of CreateSubnetGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:CreateSubnetGroup', createSubnetGroup_description - A description for the subnet group

$sel:subnetGroupName:CreateSubnetGroup', createSubnetGroup_subnetGroupName - A name for the subnet group. This value is stored as a lowercase string.

$sel:subnetIds:CreateSubnetGroup', createSubnetGroup_subnetIds - A list of VPC subnet IDs for the subnet group.

Request Lenses

createSubnetGroup_description :: Lens' CreateSubnetGroup (Maybe Text)

A description for the subnet group

createSubnetGroup_subnetGroupName :: Lens' CreateSubnetGroup Text

A name for the subnet group. This value is stored as a lowercase string.

createSubnetGroup_subnetIds :: Lens' CreateSubnetGroup [Text]

A list of VPC subnet IDs for the subnet group.

Destructuring the Response

data CreateSubnetGroupResponse

See: newCreateSubnetGroupResponse smart constructor.

Constructors

CreateSubnetGroupResponse'

Fields

subnetGroup :: Maybe SubnetGroup

Represents the output of a CreateSubnetGroup operation.

httpStatus :: Int

The response's http status code.

newCreateSubnetGroupResponse

Arguments

:: Int
$sel:httpStatus:CreateSubnetGroupResponse'

-> CreateSubnetGroupResponse

Create a value of CreateSubnetGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:subnetGroup:CreateSubnetGroupResponse', createSubnetGroupResponse_subnetGroup - Represents the output of a CreateSubnetGroup operation.

$sel:httpStatus:CreateSubnetGroupResponse', createSubnetGroupResponse_httpStatus - The response's http status code.

Response Lenses

createSubnetGroupResponse_subnetGroup :: Lens' CreateSubnetGroupResponse (Maybe SubnetGroup)

Represents the output of a CreateSubnetGroup operation.

createSubnetGroupResponse_httpStatus :: Lens' CreateSubnetGroupResponse Int

The response's http status code.