Network.AWS.DAX.DecreaseReplicationFactor

Description

Removes one or more nodes from a DAX cluster.

You cannot use DecreaseReplicationFactor to remove the last node in a DAX cluster. If you need to do this, use DeleteCluster instead.

Creating a Request

data DecreaseReplicationFactor

See: newDecreaseReplicationFactor smart constructor.

Constructors

DecreaseReplicationFactor'

Fields

nodeIdsToRemove :: Maybe [Text]

The unique identifiers of the nodes to be removed from the cluster.

availabilityZones :: Maybe [Text]

The Availability Zone(s) from which to remove nodes.

clusterName :: Text

The name of the DAX cluster from which you want to remove nodes.

newReplicationFactor' :: Int

The new number of nodes for the DAX cluster.

newDecreaseReplicationFactor

Arguments

:: Text
$sel:clusterName:DecreaseReplicationFactor'

-> Int
$sel:newReplicationFactor':DecreaseReplicationFactor'

-> DecreaseReplicationFactor

Create a value of DecreaseReplicationFactor with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nodeIdsToRemove:DecreaseReplicationFactor', decreaseReplicationFactor_nodeIdsToRemove - The unique identifiers of the nodes to be removed from the cluster.

$sel:availabilityZones:DecreaseReplicationFactor', decreaseReplicationFactor_availabilityZones - The Availability Zone(s) from which to remove nodes.

$sel:clusterName:DecreaseReplicationFactor', decreaseReplicationFactor_clusterName - The name of the DAX cluster from which you want to remove nodes.

$sel:newReplicationFactor':DecreaseReplicationFactor', decreaseReplicationFactor_newReplicationFactor - The new number of nodes for the DAX cluster.

Request Lenses

decreaseReplicationFactor_nodeIdsToRemove :: Lens' DecreaseReplicationFactor (Maybe [Text])

The unique identifiers of the nodes to be removed from the cluster.

decreaseReplicationFactor_availabilityZones :: Lens' DecreaseReplicationFactor (Maybe [Text])

The Availability Zone(s) from which to remove nodes.

decreaseReplicationFactor_clusterName :: Lens' DecreaseReplicationFactor Text

The name of the DAX cluster from which you want to remove nodes.

decreaseReplicationFactor_newReplicationFactor :: Lens' DecreaseReplicationFactor Int

The new number of nodes for the DAX cluster.

Destructuring the Response

data DecreaseReplicationFactorResponse

See: newDecreaseReplicationFactorResponse smart constructor.

Constructors

DecreaseReplicationFactorResponse'

Fields

cluster :: Maybe Cluster

A description of the DAX cluster, after you have decreased its replication factor.

httpStatus :: Int

The response's http status code.

newDecreaseReplicationFactorResponse

Arguments

:: Int
$sel:httpStatus:DecreaseReplicationFactorResponse'

-> DecreaseReplicationFactorResponse

Create a value of DecreaseReplicationFactorResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cluster:DecreaseReplicationFactorResponse', decreaseReplicationFactorResponse_cluster - A description of the DAX cluster, after you have decreased its replication factor.

$sel:httpStatus:DecreaseReplicationFactorResponse', decreaseReplicationFactorResponse_httpStatus - The response's http status code.

Response Lenses

decreaseReplicationFactorResponse_cluster :: Lens' DecreaseReplicationFactorResponse (Maybe Cluster)

A description of the DAX cluster, after you have decreased its replication factor.

decreaseReplicationFactorResponse_httpStatus :: Lens' DecreaseReplicationFactorResponse Int

The response's http status code.