Network.AWS.DAX.DescribeParameters

Description

Returns the detailed parameter list for a particular parameter group.

This operation returns paginated results.

Creating a Request

data DescribeParameters

See: newDescribeParameters smart constructor.

Constructors

DescribeParameters'

Fields

nextToken :: Maybe Text

An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults.

source :: Maybe Text

How the parameter is defined. For example, system denotes a system-defined parameter.

maxResults :: Maybe Int

The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

The value for MaxResults must be between 20 and 100.

parameterGroupName :: Text

The name of the parameter group.

newDescribeParameters

Arguments

:: Text
$sel:parameterGroupName:DescribeParameters'

-> DescribeParameters

Create a value of DescribeParameters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeParameters', describeParameters_nextToken - An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults.

$sel:source:DescribeParameters', describeParameters_source - How the parameter is defined. For example, system denotes a system-defined parameter.

$sel:maxResults:DescribeParameters', describeParameters_maxResults - The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

The value for MaxResults must be between 20 and 100.

$sel:parameterGroupName:DescribeParameters', describeParameters_parameterGroupName - The name of the parameter group.

Request Lenses

describeParameters_nextToken :: Lens' DescribeParameters (Maybe Text)

An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults.

describeParameters_source :: Lens' DescribeParameters (Maybe Text)

How the parameter is defined. For example, system denotes a system-defined parameter.

describeParameters_maxResults :: Lens' DescribeParameters (Maybe Int)

The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

The value for MaxResults must be between 20 and 100.

describeParameters_parameterGroupName :: Lens' DescribeParameters Text

The name of the parameter group.

Destructuring the Response

data DescribeParametersResponse

See: newDescribeParametersResponse smart constructor.

Constructors

DescribeParametersResponse'

Fields

nextToken :: Maybe Text

Provides an identifier to allow retrieval of paginated results.

parameters :: Maybe [Parameter]

A list of parameters within a parameter group. Each element in the list represents one parameter.

httpStatus :: Int

The response's http status code.

newDescribeParametersResponse

Arguments

:: Int
$sel:httpStatus:DescribeParametersResponse'

-> DescribeParametersResponse

Create a value of DescribeParametersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeParameters', describeParametersResponse_nextToken - Provides an identifier to allow retrieval of paginated results.

$sel:parameters:DescribeParametersResponse', describeParametersResponse_parameters - A list of parameters within a parameter group. Each element in the list represents one parameter.

$sel:httpStatus:DescribeParametersResponse', describeParametersResponse_httpStatus - The response's http status code.

Response Lenses

describeParametersResponse_nextToken :: Lens' DescribeParametersResponse (Maybe Text)

Provides an identifier to allow retrieval of paginated results.

describeParametersResponse_parameters :: Lens' DescribeParametersResponse (Maybe [Parameter])

A list of parameters within a parameter group. Each element in the list represents one parameter.

describeParametersResponse_httpStatus :: Lens' DescribeParametersResponse Int

The response's http status code.