Network.AWS.DAX.Lens

Description

Operations

DescribeParameters

describeParameters_nextToken :: Lens' DescribeParameters (Maybe Text)

An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults.

describeParameters_maxResults :: Lens' DescribeParameters (Maybe Int)

The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

The value for MaxResults must be between 20 and 100.

describeParameters_source :: Lens' DescribeParameters (Maybe Text)

How the parameter is defined. For example, system denotes a system-defined parameter.

describeParameters_parameterGroupName :: Lens' DescribeParameters Text

The name of the parameter group.

describeParametersResponse_nextToken :: Lens' DescribeParametersResponse (Maybe Text)

Provides an identifier to allow retrieval of paginated results.

describeParametersResponse_parameters :: Lens' DescribeParametersResponse (Maybe [Parameter])

A list of parameters within a parameter group. Each element in the list represents one parameter.

describeParametersResponse_httpStatus :: Lens' DescribeParametersResponse Int

The response's http status code.

DescribeClusters

describeClusters_nextToken :: Lens' DescribeClusters (Maybe Text)

An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults.

describeClusters_maxResults :: Lens' DescribeClusters (Maybe Int)

The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

The value for MaxResults must be between 20 and 100.

describeClusters_clusterNames :: Lens' DescribeClusters (Maybe [Text])

The names of the DAX clusters being described.

describeClustersResponse_nextToken :: Lens' DescribeClustersResponse (Maybe Text)

Provides an identifier to allow retrieval of paginated results.

describeClustersResponse_clusters :: Lens' DescribeClustersResponse (Maybe [Cluster])

The descriptions of your DAX clusters, in response to a DescribeClusters request.

describeClustersResponse_httpStatus :: Lens' DescribeClustersResponse Int

The response's http status code.

DescribeDefaultParameters

describeDefaultParameters_nextToken :: Lens' DescribeDefaultParameters (Maybe Text)

An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults.

describeDefaultParameters_maxResults :: Lens' DescribeDefaultParameters (Maybe Int)

The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

The value for MaxResults must be between 20 and 100.

describeDefaultParametersResponse_nextToken :: Lens' DescribeDefaultParametersResponse (Maybe Text)

Provides an identifier to allow retrieval of paginated results.

describeDefaultParametersResponse_parameters :: Lens' DescribeDefaultParametersResponse (Maybe [Parameter])

A list of parameters. Each element in the list represents one parameter.

describeDefaultParametersResponse_httpStatus :: Lens' DescribeDefaultParametersResponse Int

The response's http status code.

CreateCluster

createCluster_securityGroupIds :: Lens' CreateCluster (Maybe [Text])

A list of security group IDs to be assigned to each node in the DAX cluster. (Each of the security group ID is system-generated.)

If this parameter is not specified, DAX assigns the default VPC security group to each node.

createCluster_availabilityZones :: Lens' CreateCluster (Maybe [Text])

The Availability Zones (AZs) in which the cluster nodes will reside after the cluster has been created or updated. If provided, the length of this list must equal the ReplicationFactor parameter. If you omit this parameter, DAX will spread the nodes across Availability Zones for the highest availability.

createCluster_clusterEndpointEncryptionType :: Lens' CreateCluster (Maybe ClusterEndpointEncryptionType)

The type of encryption the cluster's endpoint should support. Values are:

NONE for no encryption

TLS for Transport Layer Security

createCluster_sSESpecification :: Lens' CreateCluster (Maybe SSESpecification)

Represents the settings used to enable server-side encryption on the cluster.

createCluster_preferredMaintenanceWindow :: Lens' CreateCluster (Maybe Text)

Specifies the weekly time range during which maintenance on the DAX cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period. Valid values for ddd are:

sun

mon

tue

wed

thu

fri

sat

Example: sun:05:00-sun:09:00

If you don't specify a preferred maintenance window when you create or modify a cache cluster, DAX assigns a 60-minute maintenance window on a randomly selected day of the week.

createCluster_tags :: Lens' CreateCluster (Maybe [Tag])

A set of tags to associate with the DAX cluster.

createCluster_parameterGroupName :: Lens' CreateCluster (Maybe Text)

The parameter group to be associated with the DAX cluster.

createCluster_notificationTopicArn :: Lens' CreateCluster (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon SNS topic to which notifications will be sent.

The Amazon SNS topic owner must be same as the DAX cluster owner.

createCluster_description :: Lens' CreateCluster (Maybe Text)

A description of the cluster.

createCluster_subnetGroupName :: Lens' CreateCluster (Maybe Text)

The name of the subnet group to be used for the replication group.

DAX clusters can only run in an Amazon VPC environment. All of the subnets that you specify in a subnet group must exist in the same VPC.

createCluster_clusterName :: Lens' CreateCluster Text

The cluster identifier. This parameter is stored as a lowercase string.

Constraints:

A name must contain from 1 to 20 alphanumeric characters or hyphens.

The first character must be a letter.

A name cannot end with a hyphen or contain two consecutive hyphens.

createCluster_nodeType :: Lens' CreateCluster Text

The compute and memory capacity of the nodes in the cluster.

createCluster_replicationFactor :: Lens' CreateCluster Int

The number of nodes in the DAX cluster. A replication factor of 1 will create a single-node cluster, without any read replicas. For additional fault tolerance, you can create a multiple node cluster with one or more read replicas. To do this, set ReplicationFactor to a number between 3 (one primary and two read replicas) and 10 (one primary and nine read replicas). If the AvailabilityZones parameter is provided, its length must equal the ReplicationFactor.

AWS recommends that you have at least two read replicas per cluster.

createCluster_iamRoleArn :: Lens' CreateCluster Text

A valid Amazon Resource Name (ARN) that identifies an IAM role. At runtime, DAX will assume this role and use the role's permissions to access DynamoDB on your behalf.

createClusterResponse_cluster :: Lens' CreateClusterResponse (Maybe Cluster)

A description of the DAX cluster that you have created.

createClusterResponse_httpStatus :: Lens' CreateClusterResponse Int

The response's http status code.

DecreaseReplicationFactor

decreaseReplicationFactor_availabilityZones :: Lens' DecreaseReplicationFactor (Maybe [Text])

The Availability Zone(s) from which to remove nodes.

decreaseReplicationFactor_nodeIdsToRemove :: Lens' DecreaseReplicationFactor (Maybe [Text])

The unique identifiers of the nodes to be removed from the cluster.

decreaseReplicationFactor_clusterName :: Lens' DecreaseReplicationFactor Text

The name of the DAX cluster from which you want to remove nodes.

decreaseReplicationFactor_newReplicationFactor :: Lens' DecreaseReplicationFactor Int

The new number of nodes for the DAX cluster.

decreaseReplicationFactorResponse_cluster :: Lens' DecreaseReplicationFactorResponse (Maybe Cluster)

A description of the DAX cluster, after you have decreased its replication factor.

decreaseReplicationFactorResponse_httpStatus :: Lens' DecreaseReplicationFactorResponse Int

The response's http status code.

UntagResource

untagResource_resourceName :: Lens' UntagResource Text

The name of the DAX resource from which the tags should be removed.

untagResource_tagKeys :: Lens' UntagResource [Text]

A list of tag keys. If the DAX cluster has any tags with these keys, then the tags are removed from the cluster.

untagResourceResponse_tags :: Lens' UntagResourceResponse (Maybe [Tag])

The tag keys that have been removed from the cluster.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

TagResource

tagResource_resourceName :: Lens' TagResource Text

The name of the DAX resource to which tags should be added.

tagResource_tags :: Lens' TagResource [Tag]

The tags to be assigned to the DAX resource.

tagResourceResponse_tags :: Lens' TagResourceResponse (Maybe [Tag])

The list of tags that are associated with the DAX resource.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

DescribeParameterGroups

describeParameterGroups_nextToken :: Lens' DescribeParameterGroups (Maybe Text)

An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults.

describeParameterGroups_maxResults :: Lens' DescribeParameterGroups (Maybe Int)

The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

The value for MaxResults must be between 20 and 100.

describeParameterGroups_parameterGroupNames :: Lens' DescribeParameterGroups (Maybe [Text])

The names of the parameter groups.

describeParameterGroupsResponse_nextToken :: Lens' DescribeParameterGroupsResponse (Maybe Text)

Provides an identifier to allow retrieval of paginated results.

describeParameterGroupsResponse_parameterGroups :: Lens' DescribeParameterGroupsResponse (Maybe [ParameterGroup])

An array of parameter groups. Each element in the array represents one parameter group.

describeParameterGroupsResponse_httpStatus :: Lens' DescribeParameterGroupsResponse Int

The response's http status code.

IncreaseReplicationFactor

increaseReplicationFactor_availabilityZones :: Lens' IncreaseReplicationFactor (Maybe [Text])

The Availability Zones (AZs) in which the cluster nodes will be created. All nodes belonging to the cluster are placed in these Availability Zones. Use this parameter if you want to distribute the nodes across multiple AZs.

increaseReplicationFactor_clusterName :: Lens' IncreaseReplicationFactor Text

The name of the DAX cluster that will receive additional nodes.

increaseReplicationFactor_newReplicationFactor :: Lens' IncreaseReplicationFactor Int

The new number of nodes for the DAX cluster.

increaseReplicationFactorResponse_cluster :: Lens' IncreaseReplicationFactorResponse (Maybe Cluster)

A description of the DAX cluster. with its new replication factor.

increaseReplicationFactorResponse_httpStatus :: Lens' IncreaseReplicationFactorResponse Int

The response's http status code.

DescribeEvents

describeEvents_nextToken :: Lens' DescribeEvents (Maybe Text)

An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults.

describeEvents_maxResults :: Lens' DescribeEvents (Maybe Int)

The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

The value for MaxResults must be between 20 and 100.

describeEvents_duration :: Lens' DescribeEvents (Maybe Int)

The number of minutes' worth of events to retrieve.

describeEvents_startTime :: Lens' DescribeEvents (Maybe UTCTime)

The beginning of the time interval to retrieve events for, specified in ISO 8601 format.

describeEvents_endTime :: Lens' DescribeEvents (Maybe UTCTime)

The end of the time interval for which to retrieve events, specified in ISO 8601 format.

describeEvents_sourceName :: Lens' DescribeEvents (Maybe Text)

The identifier of the event source for which events will be returned. If not specified, then all sources are included in the response.

describeEvents_sourceType :: Lens' DescribeEvents (Maybe SourceType)

The event source to retrieve events for. If no value is specified, all events are returned.

describeEventsResponse_nextToken :: Lens' DescribeEventsResponse (Maybe Text)

Provides an identifier to allow retrieval of paginated results.

describeEventsResponse_events :: Lens' DescribeEventsResponse (Maybe [Event])

An array of events. Each element in the array represents one event.

describeEventsResponse_httpStatus :: Lens' DescribeEventsResponse Int

The response's http status code.

UpdateParameterGroup

updateParameterGroup_parameterGroupName :: Lens' UpdateParameterGroup Text

The name of the parameter group.

updateParameterGroup_parameterNameValues :: Lens' UpdateParameterGroup [ParameterNameValue]

An array of name-value pairs for the parameters in the group. Each element in the array represents a single parameter.

record-ttl-millis and query-ttl-millis are the only supported parameter names. For more details, see Configuring TTL Settings.

updateParameterGroupResponse_parameterGroup :: Lens' UpdateParameterGroupResponse (Maybe ParameterGroup)

The parameter group that has been modified.

updateParameterGroupResponse_httpStatus :: Lens' UpdateParameterGroupResponse Int

The response's http status code.

DeleteParameterGroup

deleteParameterGroup_parameterGroupName :: Lens' DeleteParameterGroup Text

The name of the parameter group to delete.

deleteParameterGroupResponse_deletionMessage :: Lens' DeleteParameterGroupResponse (Maybe Text)

A user-specified message for this action (i.e., a reason for deleting the parameter group).

deleteParameterGroupResponse_httpStatus :: Lens' DeleteParameterGroupResponse Int

The response's http status code.

RebootNode

rebootNode_clusterName :: Lens' RebootNode Text

The name of the DAX cluster containing the node to be rebooted.

rebootNode_nodeId :: Lens' RebootNode Text

The system-assigned ID of the node to be rebooted.

rebootNodeResponse_cluster :: Lens' RebootNodeResponse (Maybe Cluster)

A description of the DAX cluster after a node has been rebooted.

rebootNodeResponse_httpStatus :: Lens' RebootNodeResponse Int

The response's http status code.

ListTags

listTags_nextToken :: Lens' ListTags (Maybe Text)

An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token.

listTags_resourceName :: Lens' ListTags Text

The name of the DAX resource to which the tags belong.

listTagsResponse_nextToken :: Lens' ListTagsResponse (Maybe Text)

If this value is present, there are additional results to be displayed. To retrieve them, call ListTags again, with NextToken set to this value.

listTagsResponse_tags :: Lens' ListTagsResponse (Maybe [Tag])

A list of tags currently associated with the DAX cluster.

listTagsResponse_httpStatus :: Lens' ListTagsResponse Int

The response's http status code.

UpdateCluster

updateCluster_securityGroupIds :: Lens' UpdateCluster (Maybe [Text])

A list of user-specified security group IDs to be assigned to each node in the DAX cluster. If this parameter is not specified, DAX assigns the default VPC security group to each node.

updateCluster_notificationTopicStatus :: Lens' UpdateCluster (Maybe Text)

The current state of the topic. A value of “active” means that notifications will be sent to the topic. A value of “inactive” means that notifications will not be sent to the topic.

updateCluster_preferredMaintenanceWindow :: Lens' UpdateCluster (Maybe Text)

A range of time when maintenance of DAX cluster software will be performed. For example: sun:01:00-sun:09:00. Cluster maintenance normally takes less than 30 minutes, and is performed automatically within the maintenance window.

updateCluster_parameterGroupName :: Lens' UpdateCluster (Maybe Text)

The name of a parameter group for this cluster.

updateCluster_notificationTopicArn :: Lens' UpdateCluster (Maybe Text)

The Amazon Resource Name (ARN) that identifies the topic.

updateCluster_description :: Lens' UpdateCluster (Maybe Text)

A description of the changes being made to the cluster.

updateCluster_clusterName :: Lens' UpdateCluster Text

The name of the DAX cluster to be modified.

updateClusterResponse_cluster :: Lens' UpdateClusterResponse (Maybe Cluster)

A description of the DAX cluster, after it has been modified.

updateClusterResponse_httpStatus :: Lens' UpdateClusterResponse Int

The response's http status code.

DeleteCluster

deleteCluster_clusterName :: Lens' DeleteCluster Text

The name of the cluster to be deleted.

deleteClusterResponse_cluster :: Lens' DeleteClusterResponse (Maybe Cluster)

A description of the DAX cluster that is being deleted.

deleteClusterResponse_httpStatus :: Lens' DeleteClusterResponse Int

The response's http status code.

CreateSubnetGroup

createSubnetGroup_description :: Lens' CreateSubnetGroup (Maybe Text)

A description for the subnet group

createSubnetGroup_subnetGroupName :: Lens' CreateSubnetGroup Text

A name for the subnet group. This value is stored as a lowercase string.

createSubnetGroup_subnetIds :: Lens' CreateSubnetGroup [Text]

A list of VPC subnet IDs for the subnet group.

createSubnetGroupResponse_subnetGroup :: Lens' CreateSubnetGroupResponse (Maybe SubnetGroup)

Represents the output of a CreateSubnetGroup operation.

createSubnetGroupResponse_httpStatus :: Lens' CreateSubnetGroupResponse Int

The response's http status code.

UpdateSubnetGroup

updateSubnetGroup_subnetIds :: Lens' UpdateSubnetGroup (Maybe [Text])

A list of subnet IDs in the subnet group.

updateSubnetGroup_description :: Lens' UpdateSubnetGroup (Maybe Text)

A description of the subnet group.

updateSubnetGroup_subnetGroupName :: Lens' UpdateSubnetGroup Text

The name of the subnet group.

updateSubnetGroupResponse_subnetGroup :: Lens' UpdateSubnetGroupResponse (Maybe SubnetGroup)

The subnet group that has been modified.

updateSubnetGroupResponse_httpStatus :: Lens' UpdateSubnetGroupResponse Int

The response's http status code.

DeleteSubnetGroup

deleteSubnetGroup_subnetGroupName :: Lens' DeleteSubnetGroup Text

The name of the subnet group to delete.

deleteSubnetGroupResponse_deletionMessage :: Lens' DeleteSubnetGroupResponse (Maybe Text)

A user-specified message for this action (i.e., a reason for deleting the subnet group).

deleteSubnetGroupResponse_httpStatus :: Lens' DeleteSubnetGroupResponse Int

The response's http status code.

DescribeSubnetGroups

describeSubnetGroups_nextToken :: Lens' DescribeSubnetGroups (Maybe Text)

An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults.

describeSubnetGroups_maxResults :: Lens' DescribeSubnetGroups (Maybe Int)

The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

The value for MaxResults must be between 20 and 100.

describeSubnetGroups_subnetGroupNames :: Lens' DescribeSubnetGroups (Maybe [Text])

The name of the subnet group.

describeSubnetGroupsResponse_subnetGroups :: Lens' DescribeSubnetGroupsResponse (Maybe [SubnetGroup])

An array of subnet groups. Each element in the array represents a single subnet group.

describeSubnetGroupsResponse_nextToken :: Lens' DescribeSubnetGroupsResponse (Maybe Text)

Provides an identifier to allow retrieval of paginated results.

describeSubnetGroupsResponse_httpStatus :: Lens' DescribeSubnetGroupsResponse Int

The response's http status code.

CreateParameterGroup

createParameterGroup_description :: Lens' CreateParameterGroup (Maybe Text)

A description of the parameter group.

createParameterGroup_parameterGroupName :: Lens' CreateParameterGroup Text

The name of the parameter group to apply to all of the clusters in this replication group.

createParameterGroupResponse_parameterGroup :: Lens' CreateParameterGroupResponse (Maybe ParameterGroup)

Represents the output of a CreateParameterGroup action.

createParameterGroupResponse_httpStatus :: Lens' CreateParameterGroupResponse Int

The response's http status code.

Types

Cluster

cluster_clusterArn :: Lens' Cluster (Maybe Text)

The Amazon Resource Name (ARN) that uniquely identifies the cluster.

cluster_status :: Lens' Cluster (Maybe Text)

The current status of the cluster.

cluster_iamRoleArn :: Lens' Cluster (Maybe Text)

A valid Amazon Resource Name (ARN) that identifies an IAM role. At runtime, DAX will assume this role and use the role's permissions to access DynamoDB on your behalf.

cluster_subnetGroup :: Lens' Cluster (Maybe Text)

The subnet group where the DAX cluster is running.

cluster_totalNodes :: Lens' Cluster (Maybe Int)

The total number of nodes in the cluster.

cluster_parameterGroup :: Lens' Cluster (Maybe ParameterGroupStatus)

The parameter group being used by nodes in the cluster.

cluster_clusterEndpointEncryptionType :: Lens' Cluster (Maybe ClusterEndpointEncryptionType)

The type of encryption supported by the cluster's endpoint. Values are:

NONE for no encryption

TLS for Transport Layer Security

cluster_nodes :: Lens' Cluster (Maybe [Node])

A list of nodes that are currently in the cluster.

cluster_notificationConfiguration :: Lens' Cluster (Maybe NotificationConfiguration)

Describes a notification topic and its status. Notification topics are used for publishing DAX events to subscribers using Amazon Simple Notification Service (SNS).

cluster_securityGroups :: Lens' Cluster (Maybe [SecurityGroupMembership])

A list of security groups, and the status of each, for the nodes in the cluster.

cluster_activeNodes :: Lens' Cluster (Maybe Int)

The number of nodes in the cluster that are active (i.e., capable of serving requests).

cluster_preferredMaintenanceWindow :: Lens' Cluster (Maybe Text)

A range of time when maintenance of DAX cluster software will be performed. For example: sun:01:00-sun:09:00. Cluster maintenance normally takes less than 30 minutes, and is performed automatically within the maintenance window.

cluster_description :: Lens' Cluster (Maybe Text)

The description of the cluster.

cluster_sSEDescription :: Lens' Cluster (Maybe SSEDescription)

The description of the server-side encryption status on the specified DAX cluster.

cluster_nodeIdsToRemove :: Lens' Cluster (Maybe [Text])

A list of nodes to be removed from the cluster.

cluster_clusterDiscoveryEndpoint :: Lens' Cluster (Maybe Endpoint)

The endpoint for this DAX cluster, consisting of a DNS name, a port number, and a URL. Applications should use the URL to configure the DAX client to find their cluster.

cluster_nodeType :: Lens' Cluster (Maybe Text)

The node type for the nodes in the cluster. (All nodes in a DAX cluster are of the same type.)

cluster_clusterName :: Lens' Cluster (Maybe Text)

The name of the DAX cluster.

Endpoint

endpoint_address :: Lens' Endpoint (Maybe Text)

The DNS hostname of the endpoint.

endpoint_port :: Lens' Endpoint (Maybe Int)

The port number that applications should use to connect to the endpoint.

endpoint_url :: Lens' Endpoint (Maybe Text)

The URL that applications should use to connect to the endpoint. The default ports are 8111 for the "dax" protocol and 9111 for the "daxs" protocol.

Event

event_message :: Lens' Event (Maybe Text)

A user-defined message associated with the event.

event_sourceName :: Lens' Event (Maybe Text)

The source of the event. For example, if the event occurred at the node level, the source would be the node ID.

event_date :: Lens' Event (Maybe UTCTime)

The date and time when the event occurred.

event_sourceType :: Lens' Event (Maybe SourceType)

Specifies the origin of this event - a cluster, a parameter group, a node ID, etc.

Node

node_nodeStatus :: Lens' Node (Maybe Text)

The current status of the node. For example: available.

node_nodeId :: Lens' Node (Maybe Text)

A system-generated identifier for the node.

node_parameterGroupStatus :: Lens' Node (Maybe Text)

The status of the parameter group associated with this node. For example, in-sync.

node_availabilityZone :: Lens' Node (Maybe Text)

The Availability Zone (AZ) in which the node has been deployed.

node_nodeCreateTime :: Lens' Node (Maybe UTCTime)

The date and time (in UNIX epoch format) when the node was launched.

node_endpoint :: Lens' Node (Maybe Endpoint)

The endpoint for the node, consisting of a DNS name and a port number. Client applications can connect directly to a node endpoint, if desired (as an alternative to allowing DAX client software to intelligently route requests and responses to nodes in the DAX cluster.

NodeTypeSpecificValue

nodeTypeSpecificValue_value :: Lens' NodeTypeSpecificValue (Maybe Text)

The parameter value for this node type.

nodeTypeSpecificValue_nodeType :: Lens' NodeTypeSpecificValue (Maybe Text)

A node type to which the parameter value applies.

NotificationConfiguration

notificationConfiguration_topicStatus :: Lens' NotificationConfiguration (Maybe Text)

The current state of the topic. A value of “active” means that notifications will be sent to the topic. A value of “inactive” means that notifications will not be sent to the topic.

notificationConfiguration_topicArn :: Lens' NotificationConfiguration (Maybe Text)

The Amazon Resource Name (ARN) that identifies the topic.

Parameter

parameter_changeType :: Lens' Parameter (Maybe ChangeType)

The conditions under which changes to this parameter can be applied. For example, requires-reboot indicates that a new value for this parameter will only take effect if a node is rebooted.

parameter_allowedValues :: Lens' Parameter (Maybe Text)

A range of values within which the parameter can be set.

parameter_source :: Lens' Parameter (Maybe Text)

How the parameter is defined. For example, system denotes a system-defined parameter.

parameter_parameterValue :: Lens' Parameter (Maybe Text)

The value for the parameter.

parameter_parameterType :: Lens' Parameter (Maybe ParameterType)

Determines whether the parameter can be applied to any nodes, or only nodes of a particular type.

parameter_parameterName :: Lens' Parameter (Maybe Text)

The name of the parameter.

parameter_description :: Lens' Parameter (Maybe Text)

A description of the parameter

parameter_nodeTypeSpecificValues :: Lens' Parameter (Maybe [NodeTypeSpecificValue])

A list of node types, and specific parameter values for each node.

parameter_dataType :: Lens' Parameter (Maybe Text)

The data type of the parameter. For example, integer:

parameter_isModifiable :: Lens' Parameter (Maybe IsModifiable)

Whether the customer is allowed to modify the parameter.

ParameterGroup

parameterGroup_parameterGroupName :: Lens' ParameterGroup (Maybe Text)

The name of the parameter group.

parameterGroup_description :: Lens' ParameterGroup (Maybe Text)

A description of the parameter group.

ParameterGroupStatus

parameterGroupStatus_nodeIdsToReboot :: Lens' ParameterGroupStatus (Maybe [Text])

The node IDs of one or more nodes to be rebooted.

parameterGroupStatus_parameterGroupName :: Lens' ParameterGroupStatus (Maybe Text)

The name of the parameter group.

parameterGroupStatus_parameterApplyStatus :: Lens' ParameterGroupStatus (Maybe Text)

The status of parameter updates.

ParameterNameValue

parameterNameValue_parameterValue :: Lens' ParameterNameValue (Maybe Text)

The value of the parameter.

parameterNameValue_parameterName :: Lens' ParameterNameValue (Maybe Text)

The name of the parameter.

SSEDescription

sSEDescription_status :: Lens' SSEDescription (Maybe SSEStatus)

The current state of server-side encryption:

ENABLING - Server-side encryption is being enabled.

ENABLED - Server-side encryption is enabled.

DISABLING - Server-side encryption is being disabled.

DISABLED - Server-side encryption is disabled.

SSESpecification

sSESpecification_enabled :: Lens' SSESpecification Bool

Indicates whether server-side encryption is enabled (true) or disabled (false) on the cluster.

SecurityGroupMembership

securityGroupMembership_status :: Lens' SecurityGroupMembership (Maybe Text)

The status of this security group.

securityGroupMembership_securityGroupIdentifier :: Lens' SecurityGroupMembership (Maybe Text)

The unique ID for this security group.

Subnet

subnet_subnetIdentifier :: Lens' Subnet (Maybe Text)

The system-assigned identifier for the subnet.

subnet_subnetAvailabilityZone :: Lens' Subnet (Maybe Text)

The Availability Zone (AZ) for the subnet.

SubnetGroup

subnetGroup_description :: Lens' SubnetGroup (Maybe Text)

The description of the subnet group.

subnetGroup_subnetGroupName :: Lens' SubnetGroup (Maybe Text)

The name of the subnet group.

subnetGroup_vpcId :: Lens' SubnetGroup (Maybe Text)

The Amazon Virtual Private Cloud identifier (VPC ID) of the subnet group.

subnetGroup_subnets :: Lens' SubnetGroup (Maybe [Subnet])

A list of subnets associated with the subnet group.

Tag

tag_key :: Lens' Tag (Maybe Text)

The key for the tag. Tag keys are case sensitive. Every DAX cluster can only have one tag with the same key. If you try to add an existing tag (same key), the existing tag value will be updated to the new value.

tag_value :: Lens' Tag (Maybe Text)

The value of the tag. Tag values are case-sensitive and can be null.