Network.AWS.DAX.RebootNode

Description

Reboots a single node of a DAX cluster. The reboot action takes place as soon as possible. During the reboot, the node status is set to REBOOTING.

RebootNode restarts the DAX engine process and does not remove the contents of the cache.

Creating a Request

data RebootNode

See: newRebootNode smart constructor.

Constructors

RebootNode'

Fields

clusterName :: Text

The name of the DAX cluster containing the node to be rebooted.

nodeId :: Text

The system-assigned ID of the node to be rebooted.

newRebootNode

Arguments

:: Text
$sel:clusterName:RebootNode'

-> Text
$sel:nodeId:RebootNode'

-> RebootNode

Create a value of RebootNode with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clusterName:RebootNode', rebootNode_clusterName - The name of the DAX cluster containing the node to be rebooted.

$sel:nodeId:RebootNode', rebootNode_nodeId - The system-assigned ID of the node to be rebooted.

Request Lenses

rebootNode_clusterName :: Lens' RebootNode Text

The name of the DAX cluster containing the node to be rebooted.

rebootNode_nodeId :: Lens' RebootNode Text

The system-assigned ID of the node to be rebooted.

Destructuring the Response

data RebootNodeResponse

See: newRebootNodeResponse smart constructor.

Constructors

RebootNodeResponse'

Fields

cluster :: Maybe Cluster

A description of the DAX cluster after a node has been rebooted.

httpStatus :: Int

The response's http status code.

newRebootNodeResponse

Arguments

:: Int
$sel:httpStatus:RebootNodeResponse'

-> RebootNodeResponse

Create a value of RebootNodeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cluster:RebootNodeResponse', rebootNodeResponse_cluster - A description of the DAX cluster after a node has been rebooted.

$sel:httpStatus:RebootNodeResponse', rebootNodeResponse_httpStatus - The response's http status code.

Response Lenses

rebootNodeResponse_cluster :: Lens' RebootNodeResponse (Maybe Cluster)

A description of the DAX cluster after a node has been rebooted.

rebootNodeResponse_httpStatus :: Lens' RebootNodeResponse Int

The response's http status code.