Network.AWS.DAX.Types.Node

Description

Documentation

data Node

Represents an individual node within a DAX cluster.

See: newNode smart constructor.

Constructors

Node'

Fields

nodeStatus :: Maybe Text

The current status of the node. For example: available.

nodeId :: Maybe Text

A system-generated identifier for the node.

parameterGroupStatus :: Maybe Text

The status of the parameter group associated with this node. For example, in-sync.

availabilityZone :: Maybe Text

The Availability Zone (AZ) in which the node has been deployed.

nodeCreateTime :: Maybe POSIX

The date and time (in UNIX epoch format) when the node was launched.

endpoint :: Maybe Endpoint

The endpoint for the node, consisting of a DNS name and a port number. Client applications can connect directly to a node endpoint, if desired (as an alternative to allowing DAX client software to intelligently route requests and responses to nodes in the DAX cluster.

newNode :: Node

Create a value of Node with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nodeStatus:Node', node_nodeStatus - The current status of the node. For example: available.

$sel:nodeId:Node', node_nodeId - A system-generated identifier for the node.

$sel:parameterGroupStatus:Node', node_parameterGroupStatus - The status of the parameter group associated with this node. For example, in-sync.

$sel:availabilityZone:Node', node_availabilityZone - The Availability Zone (AZ) in which the node has been deployed.

$sel:nodeCreateTime:Node', node_nodeCreateTime - The date and time (in UNIX epoch format) when the node was launched.

$sel:endpoint:Node', node_endpoint - The endpoint for the node, consisting of a DNS name and a port number. Client applications can connect directly to a node endpoint, if desired (as an alternative to allowing DAX client software to intelligently route requests and responses to nodes in the DAX cluster.

node_nodeStatus :: Lens' Node (Maybe Text)

The current status of the node. For example: available.

node_nodeId :: Lens' Node (Maybe Text)

A system-generated identifier for the node.

node_parameterGroupStatus :: Lens' Node (Maybe Text)

The status of the parameter group associated with this node. For example, in-sync.

node_availabilityZone :: Lens' Node (Maybe Text)

The Availability Zone (AZ) in which the node has been deployed.

node_nodeCreateTime :: Lens' Node (Maybe UTCTime)

The date and time (in UNIX epoch format) when the node was launched.

node_endpoint :: Lens' Node (Maybe Endpoint)

The endpoint for the node, consisting of a DNS name and a port number. Client applications can connect directly to a node endpoint, if desired (as an alternative to allowing DAX client software to intelligently route requests and responses to nodes in the DAX cluster.