Network.AWS.DAX.UpdateCluster

Description

Modifies the settings for a DAX cluster. You can use this action to change one or more cluster configuration parameters by specifying the parameters and the new values.

Creating a Request

data UpdateCluster

See: newUpdateCluster smart constructor.

Constructors

UpdateCluster'

Fields

securityGroupIds :: Maybe [Text]

A list of user-specified security group IDs to be assigned to each node in the DAX cluster. If this parameter is not specified, DAX assigns the default VPC security group to each node.

preferredMaintenanceWindow :: Maybe Text

A range of time when maintenance of DAX cluster software will be performed. For example: sun:01:00-sun:09:00. Cluster maintenance normally takes less than 30 minutes, and is performed automatically within the maintenance window.

notificationTopicStatus :: Maybe Text

The current state of the topic. A value of “active” means that notifications will be sent to the topic. A value of “inactive” means that notifications will not be sent to the topic.

description :: Maybe Text

A description of the changes being made to the cluster.

notificationTopicArn :: Maybe Text

The Amazon Resource Name (ARN) that identifies the topic.

parameterGroupName :: Maybe Text

The name of a parameter group for this cluster.

clusterName :: Text

The name of the DAX cluster to be modified.

newUpdateCluster

Arguments

:: Text
$sel:clusterName:UpdateCluster'

-> UpdateCluster

Create a value of UpdateCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:securityGroupIds:UpdateCluster', updateCluster_securityGroupIds - A list of user-specified security group IDs to be assigned to each node in the DAX cluster. If this parameter is not specified, DAX assigns the default VPC security group to each node.

$sel:preferredMaintenanceWindow:UpdateCluster', updateCluster_preferredMaintenanceWindow - A range of time when maintenance of DAX cluster software will be performed. For example: sun:01:00-sun:09:00. Cluster maintenance normally takes less than 30 minutes, and is performed automatically within the maintenance window.

$sel:notificationTopicStatus:UpdateCluster', updateCluster_notificationTopicStatus - The current state of the topic. A value of “active” means that notifications will be sent to the topic. A value of “inactive” means that notifications will not be sent to the topic.

$sel:description:UpdateCluster', updateCluster_description - A description of the changes being made to the cluster.

$sel:notificationTopicArn:UpdateCluster', updateCluster_notificationTopicArn - The Amazon Resource Name (ARN) that identifies the topic.

$sel:parameterGroupName:UpdateCluster', updateCluster_parameterGroupName - The name of a parameter group for this cluster.

$sel:clusterName:UpdateCluster', updateCluster_clusterName - The name of the DAX cluster to be modified.

Request Lenses

updateCluster_securityGroupIds :: Lens' UpdateCluster (Maybe [Text])

A list of user-specified security group IDs to be assigned to each node in the DAX cluster. If this parameter is not specified, DAX assigns the default VPC security group to each node.

updateCluster_preferredMaintenanceWindow :: Lens' UpdateCluster (Maybe Text)

A range of time when maintenance of DAX cluster software will be performed. For example: sun:01:00-sun:09:00. Cluster maintenance normally takes less than 30 minutes, and is performed automatically within the maintenance window.

updateCluster_notificationTopicStatus :: Lens' UpdateCluster (Maybe Text)

The current state of the topic. A value of “active” means that notifications will be sent to the topic. A value of “inactive” means that notifications will not be sent to the topic.

updateCluster_description :: Lens' UpdateCluster (Maybe Text)

A description of the changes being made to the cluster.

updateCluster_notificationTopicArn :: Lens' UpdateCluster (Maybe Text)

The Amazon Resource Name (ARN) that identifies the topic.

updateCluster_parameterGroupName :: Lens' UpdateCluster (Maybe Text)

The name of a parameter group for this cluster.

updateCluster_clusterName :: Lens' UpdateCluster Text

The name of the DAX cluster to be modified.

Destructuring the Response

data UpdateClusterResponse

See: newUpdateClusterResponse smart constructor.

Constructors

UpdateClusterResponse'

Fields

cluster :: Maybe Cluster

A description of the DAX cluster, after it has been modified.

httpStatus :: Int

The response's http status code.

newUpdateClusterResponse

Arguments

:: Int
$sel:httpStatus:UpdateClusterResponse'

-> UpdateClusterResponse

Create a value of UpdateClusterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cluster:UpdateClusterResponse', updateClusterResponse_cluster - A description of the DAX cluster, after it has been modified.

$sel:httpStatus:UpdateClusterResponse', updateClusterResponse_httpStatus - The response's http status code.

Response Lenses

updateClusterResponse_cluster :: Lens' UpdateClusterResponse (Maybe Cluster)

A description of the DAX cluster, after it has been modified.

updateClusterResponse_httpStatus :: Lens' UpdateClusterResponse Int

The response's http status code.