Network.AWS.DLM

Description

Derived from API version 2018-01-12 of the AWS service descriptions, licensed under Apache 2.0.

Amazon Data Lifecycle Manager

With Amazon Data Lifecycle Manager, you can manage the lifecycle of your Amazon Web Services resources. You create lifecycle policies, which are used to automate operations on the specified resources.

Amazon DLM supports Amazon EBS volumes and snapshots. For information about using Amazon DLM with Amazon EBS, see Automating the Amazon EBS Snapshot Lifecycle in the Amazon EC2 User Guide.

Service Configuration

defaultService :: Service

API version 2018-01-12 of the Amazon Data Lifecycle Manager SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by DLM.

InvalidRequestException

_InvalidRequestException :: AsError a => Getting (First ServiceError) a ServiceError

Bad request. The request is missing required parameters or has invalid parameters.

InternalServerException

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

The service failed in an unexpected way.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

A requested resource was not found.

LimitExceededException

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The request failed because a limit was exceeded.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

DeleteLifecyclePolicy

data DeleteLifecyclePolicy

See: newDeleteLifecyclePolicy smart constructor.

Constructors

DeleteLifecyclePolicy' Text

newDeleteLifecyclePolicy

Arguments

:: Text
$sel:policyId:DeleteLifecyclePolicy'

-> DeleteLifecyclePolicy

Create a value of DeleteLifecyclePolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policyId:DeleteLifecyclePolicy', deleteLifecyclePolicy_policyId - The identifier of the lifecycle policy.

data DeleteLifecyclePolicyResponse

See: newDeleteLifecyclePolicyResponse smart constructor.

Constructors

DeleteLifecyclePolicyResponse' Int

newDeleteLifecyclePolicyResponse

Arguments

:: Int
$sel:httpStatus:DeleteLifecyclePolicyResponse'

-> DeleteLifecyclePolicyResponse

Create a value of DeleteLifecyclePolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteLifecyclePolicyResponse', deleteLifecyclePolicyResponse_httpStatus - The response's http status code.

UpdateLifecyclePolicy

data UpdateLifecyclePolicy

See: newUpdateLifecyclePolicy smart constructor.

Constructors

UpdateLifecyclePolicy' (Maybe SettablePolicyStateValues) (Maybe PolicyDetails) (Maybe Text) (Maybe Text) Text

newUpdateLifecyclePolicy

Arguments

:: Text
$sel:policyId:UpdateLifecyclePolicy'

-> UpdateLifecyclePolicy

Create a value of UpdateLifecyclePolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:UpdateLifecyclePolicy', updateLifecyclePolicy_state - The desired activation state of the lifecycle policy after creation.

$sel:policyDetails:UpdateLifecyclePolicy', updateLifecyclePolicy_policyDetails - The configuration of the lifecycle policy. You cannot update the policy type or the resource type.

$sel:executionRoleArn:UpdateLifecyclePolicy', updateLifecyclePolicy_executionRoleArn - The Amazon Resource Name (ARN) of the IAM role used to run the operations specified by the lifecycle policy.

$sel:description:UpdateLifecyclePolicy', updateLifecyclePolicy_description - A description of the lifecycle policy.

$sel:policyId:UpdateLifecyclePolicy', updateLifecyclePolicy_policyId - The identifier of the lifecycle policy.

data UpdateLifecyclePolicyResponse

See: newUpdateLifecyclePolicyResponse smart constructor.

Constructors

UpdateLifecyclePolicyResponse' Int

newUpdateLifecyclePolicyResponse

Arguments

:: Int
$sel:httpStatus:UpdateLifecyclePolicyResponse'

-> UpdateLifecyclePolicyResponse

Create a value of UpdateLifecyclePolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateLifecyclePolicyResponse', updateLifecyclePolicyResponse_httpStatus - The response's http status code.

ListTagsForResource

data ListTagsForResource

See: newListTagsForResource smart constructor.

Constructors

ListTagsForResource' Text

newListTagsForResource

Arguments

:: Text
$sel:resourceArn:ListTagsForResource'

-> ListTagsForResource

Create a value of ListTagsForResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:ListTagsForResource', listTagsForResource_resourceArn - The Amazon Resource Name (ARN) of the resource.

data ListTagsForResourceResponse

See: newListTagsForResourceResponse smart constructor.

Constructors

ListTagsForResourceResponse' (Maybe (HashMap Text Text)) Int

newListTagsForResourceResponse

Arguments

:: Int
$sel:httpStatus:ListTagsForResourceResponse'

-> ListTagsForResourceResponse

Create a value of ListTagsForResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:ListTagsForResourceResponse', listTagsForResourceResponse_tags - Information about the tags.

$sel:httpStatus:ListTagsForResourceResponse', listTagsForResourceResponse_httpStatus - The response's http status code.

CreateLifecyclePolicy

data CreateLifecyclePolicy

See: newCreateLifecyclePolicy smart constructor.

Constructors

CreateLifecyclePolicy' (Maybe (HashMap Text Text)) Text Text SettablePolicyStateValues PolicyDetails

newCreateLifecyclePolicy

Arguments

:: Text
$sel:executionRoleArn:CreateLifecyclePolicy'

-> Text
$sel:description:CreateLifecyclePolicy'

-> SettablePolicyStateValues
$sel:state:CreateLifecyclePolicy'

-> PolicyDetails
$sel:policyDetails:CreateLifecyclePolicy'

-> CreateLifecyclePolicy

Create a value of CreateLifecyclePolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateLifecyclePolicy', createLifecyclePolicy_tags - The tags to apply to the lifecycle policy during creation.

$sel:executionRoleArn:CreateLifecyclePolicy', createLifecyclePolicy_executionRoleArn - The Amazon Resource Name (ARN) of the IAM role used to run the operations specified by the lifecycle policy.

$sel:description:CreateLifecyclePolicy', createLifecyclePolicy_description - A description of the lifecycle policy. The characters ^[0-9A-Za-z _-]+$ are supported.

$sel:state:CreateLifecyclePolicy', createLifecyclePolicy_state - The desired activation state of the lifecycle policy after creation.

$sel:policyDetails:CreateLifecyclePolicy', createLifecyclePolicy_policyDetails - The configuration details of the lifecycle policy.

data CreateLifecyclePolicyResponse

See: newCreateLifecyclePolicyResponse smart constructor.

Constructors

CreateLifecyclePolicyResponse' (Maybe Text) Int

newCreateLifecyclePolicyResponse

Arguments

:: Int
$sel:httpStatus:CreateLifecyclePolicyResponse'

-> CreateLifecyclePolicyResponse

Create a value of CreateLifecyclePolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policyId:CreateLifecyclePolicyResponse', createLifecyclePolicyResponse_policyId - The identifier of the lifecycle policy.

$sel:httpStatus:CreateLifecyclePolicyResponse', createLifecyclePolicyResponse_httpStatus - The response's http status code.

GetLifecyclePolicy

data GetLifecyclePolicy

See: newGetLifecyclePolicy smart constructor.

Constructors

GetLifecyclePolicy' Text

newGetLifecyclePolicy

Arguments

:: Text
$sel:policyId:GetLifecyclePolicy'

-> GetLifecyclePolicy

Create a value of GetLifecyclePolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policyId:GetLifecyclePolicy', getLifecyclePolicy_policyId - The identifier of the lifecycle policy.

data GetLifecyclePolicyResponse

See: newGetLifecyclePolicyResponse smart constructor.

Constructors

GetLifecyclePolicyResponse' (Maybe LifecyclePolicy) Int

newGetLifecyclePolicyResponse

Arguments

:: Int
$sel:httpStatus:GetLifecyclePolicyResponse'

-> GetLifecyclePolicyResponse

Create a value of GetLifecyclePolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policy:GetLifecyclePolicyResponse', getLifecyclePolicyResponse_policy - Detailed information about the lifecycle policy.

$sel:httpStatus:GetLifecyclePolicyResponse', getLifecyclePolicyResponse_httpStatus - The response's http status code.

TagResource

data TagResource

See: newTagResource smart constructor.

Constructors

TagResource' Text (HashMap Text Text)

newTagResource

Arguments

:: Text
$sel:resourceArn:TagResource'

-> TagResource

Create a value of TagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:TagResource', tagResource_resourceArn - The Amazon Resource Name (ARN) of the resource.

$sel:tags:TagResource', tagResource_tags - One or more tags.

data TagResourceResponse

See: newTagResourceResponse smart constructor.

Constructors

TagResourceResponse' Int

newTagResourceResponse

Arguments

:: Int
$sel:httpStatus:TagResourceResponse'

-> TagResourceResponse

Create a value of TagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:TagResourceResponse', tagResourceResponse_httpStatus - The response's http status code.

UntagResource

data UntagResource

See: newUntagResource smart constructor.

Constructors

UntagResource' Text (NonEmpty Text)

newUntagResource

Arguments

:: Text
$sel:resourceArn:UntagResource'

-> NonEmpty Text
$sel:tagKeys:UntagResource'

-> UntagResource

Create a value of UntagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:UntagResource', untagResource_resourceArn - The Amazon Resource Name (ARN) of the resource.

$sel:tagKeys:UntagResource', untagResource_tagKeys - The tag keys.

data UntagResourceResponse

See: newUntagResourceResponse smart constructor.

Constructors

UntagResourceResponse' Int

newUntagResourceResponse

Arguments

:: Int
$sel:httpStatus:UntagResourceResponse'

-> UntagResourceResponse

Create a value of UntagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UntagResourceResponse', untagResourceResponse_httpStatus - The response's http status code.

GetLifecyclePolicies

data GetLifecyclePolicies

See: newGetLifecyclePolicies smart constructor.

Constructors

GetLifecyclePolicies' (Maybe GettablePolicyStateValues) (Maybe (NonEmpty Text)) (Maybe [Text]) (Maybe [Text]) (Maybe (NonEmpty ResourceTypeValues))

newGetLifecyclePolicies :: GetLifecyclePolicies

Create a value of GetLifecyclePolicies with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:GetLifecyclePolicies', getLifecyclePolicies_state - The activation state.

$sel:targetTags:GetLifecyclePolicies', getLifecyclePolicies_targetTags - The target tag for a policy.

Tags are strings in the format key=value.

$sel:tagsToAdd:GetLifecyclePolicies', getLifecyclePolicies_tagsToAdd - The tags to add to objects created by the policy.

Tags are strings in the format key=value.

These user-defined tags are added in addition to the Amazon Web Services-added lifecycle tags.

$sel:policyIds:GetLifecyclePolicies', getLifecyclePolicies_policyIds - The identifiers of the data lifecycle policies.

$sel:resourceTypes:GetLifecyclePolicies', getLifecyclePolicies_resourceTypes - The resource type.

data GetLifecyclePoliciesResponse

See: newGetLifecyclePoliciesResponse smart constructor.

Constructors

GetLifecyclePoliciesResponse' (Maybe [LifecyclePolicySummary]) Int

newGetLifecyclePoliciesResponse

Arguments

:: Int
$sel:httpStatus:GetLifecyclePoliciesResponse'

-> GetLifecyclePoliciesResponse

Create a value of GetLifecyclePoliciesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policies:GetLifecyclePoliciesResponse', getLifecyclePoliciesResponse_policies - Summary information about the lifecycle policies.

$sel:httpStatus:GetLifecyclePoliciesResponse', getLifecyclePoliciesResponse_httpStatus - The response's http status code.

Types

EventSourceValues

newtype EventSourceValues

Constructors

EventSourceValues'

Fields

fromEventSourceValues :: Text

Bundled Patterns

pattern EventSourceValues_MANAGED_CWE :: EventSourceValues

EventTypeValues

newtype EventTypeValues

Constructors

EventTypeValues'

Fields

fromEventTypeValues :: Text

Bundled Patterns

pattern EventTypeValues_ShareSnapshot :: EventTypeValues

GettablePolicyStateValues

newtype GettablePolicyStateValues

Constructors

GettablePolicyStateValues'

Fields

fromGettablePolicyStateValues :: Text

Bundled Patterns

pattern GettablePolicyStateValues_DISABLED :: GettablePolicyStateValues

pattern GettablePolicyStateValues_ENABLED :: GettablePolicyStateValues

pattern GettablePolicyStateValues_ERROR :: GettablePolicyStateValues

IntervalUnitValues

newtype IntervalUnitValues

Constructors

IntervalUnitValues'

Fields

fromIntervalUnitValues :: Text

Bundled Patterns

pattern IntervalUnitValues_HOURS :: IntervalUnitValues

LocationValues

newtype LocationValues

Constructors

LocationValues'

Fields

fromLocationValues :: Text

Bundled Patterns

pattern LocationValues_CLOUD :: LocationValues

pattern LocationValues_OUTPOST_LOCAL :: LocationValues

PolicyTypeValues

newtype PolicyTypeValues

Constructors

PolicyTypeValues'

Fields

fromPolicyTypeValues :: Text

Bundled Patterns

pattern PolicyTypeValues_EBS_SNAPSHOT_MANAGEMENT :: PolicyTypeValues

pattern PolicyTypeValues_EVENT_BASED_POLICY :: PolicyTypeValues

pattern PolicyTypeValues_IMAGE_MANAGEMENT :: PolicyTypeValues

ResourceLocationValues

newtype ResourceLocationValues

Constructors

ResourceLocationValues'

Fields

fromResourceLocationValues :: Text

Bundled Patterns

pattern ResourceLocationValues_CLOUD :: ResourceLocationValues

pattern ResourceLocationValues_OUTPOST :: ResourceLocationValues

ResourceTypeValues

newtype ResourceTypeValues

Constructors

ResourceTypeValues'

Fields

fromResourceTypeValues :: Text

Bundled Patterns

pattern ResourceTypeValues_INSTANCE :: ResourceTypeValues

pattern ResourceTypeValues_VOLUME :: ResourceTypeValues

RetentionIntervalUnitValues

newtype RetentionIntervalUnitValues

Constructors

RetentionIntervalUnitValues'

Fields

fromRetentionIntervalUnitValues :: Text

Bundled Patterns

pattern RetentionIntervalUnitValues_DAYS :: RetentionIntervalUnitValues

pattern RetentionIntervalUnitValues_MONTHS :: RetentionIntervalUnitValues

pattern RetentionIntervalUnitValues_WEEKS :: RetentionIntervalUnitValues

pattern RetentionIntervalUnitValues_YEARS :: RetentionIntervalUnitValues

SettablePolicyStateValues

newtype SettablePolicyStateValues

Constructors

SettablePolicyStateValues'

Fields

fromSettablePolicyStateValues :: Text

Bundled Patterns

pattern SettablePolicyStateValues_DISABLED :: SettablePolicyStateValues

pattern SettablePolicyStateValues_ENABLED :: SettablePolicyStateValues

Action

data Action

Specifies an action for an event-based policy.

See: newAction smart constructor.

Constructors

Action' Text [CrossRegionCopyAction]

newAction

Arguments

:: Text
$sel:name:Action'

-> Action

Create a value of Action with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:Action', action_name - A descriptive name for the action.

$sel:crossRegionCopy:Action', action_crossRegionCopy - The rule for copying shared snapshots across Regions.

CreateRule

data CreateRule

Specifies when to create snapshots of EBS volumes.

You must specify either a Cron expression or an interval, interval unit, and start time. You cannot specify both.

See: newCreateRule smart constructor.

Constructors

CreateRule' (Maybe LocationValues) (Maybe Natural) (Maybe Text) (Maybe [Text]) (Maybe IntervalUnitValues)

newCreateRule :: CreateRule

Create a value of CreateRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:location:CreateRule', createRule_location - Specifies the destination for snapshots created by the policy. To create snapshots in the same Region as the source resource, specify CLOUD. To create snapshots on the same Outpost as the source resource, specify OUTPOST_LOCAL. If you omit this parameter, CLOUD is used by default.

If the policy targets resources in an Amazon Web Services Region, then you must create snapshots in the same Region as the source resource.

If the policy targets resources on an Outpost, then you can create snapshots on the same Outpost as the source resource, or in the Region of that Outpost.

$sel:interval:CreateRule', createRule_interval - The interval between snapshots. The supported values are 1, 2, 3, 4, 6, 8, 12, and 24.

$sel:cronExpression:CreateRule', createRule_cronExpression - The schedule, as a Cron expression. The schedule interval must be between 1 hour and 1 year. For more information, see Cron expressions in the Amazon CloudWatch User Guide.

$sel:times:CreateRule', createRule_times - The time, in UTC, to start the operation. The supported format is hh:mm.

The operation occurs within a one-hour window following the specified time. If you do not specify a time, Amazon DLM selects a time within the next 24 hours.

$sel:intervalUnit:CreateRule', createRule_intervalUnit - The interval unit.

CrossRegionCopyAction

data CrossRegionCopyAction

Specifies a rule for copying shared snapshots across Regions.

See: newCrossRegionCopyAction smart constructor.

Constructors

CrossRegionCopyAction' (Maybe CrossRegionCopyRetainRule) Text EncryptionConfiguration

newCrossRegionCopyAction

Arguments

:: Text
$sel:target:CrossRegionCopyAction'

-> EncryptionConfiguration
$sel:encryptionConfiguration:CrossRegionCopyAction'

-> CrossRegionCopyAction

Create a value of CrossRegionCopyAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:retainRule:CrossRegionCopyAction', crossRegionCopyAction_retainRule - Undocumented member.

$sel:target:CrossRegionCopyAction', crossRegionCopyAction_target - The target Region.

$sel:encryptionConfiguration:CrossRegionCopyAction', crossRegionCopyAction_encryptionConfiguration - The encryption settings for the copied snapshot.

CrossRegionCopyDeprecateRule

data CrossRegionCopyDeprecateRule

Specifies an AMI deprecation rule for cross-Region AMI copies created by a cross-Region copy rule.

See: newCrossRegionCopyDeprecateRule smart constructor.

Constructors

CrossRegionCopyDeprecateRule' (Maybe Natural) (Maybe RetentionIntervalUnitValues)

newCrossRegionCopyDeprecateRule :: CrossRegionCopyDeprecateRule

Create a value of CrossRegionCopyDeprecateRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:interval:CrossRegionCopyDeprecateRule', crossRegionCopyDeprecateRule_interval - The period after which to deprecate the cross-Region AMI copies. The period must be less than or equal to the cross-Region AMI copy retention period, and it can't be greater than 10 years. This is equivalent to 120 months, 520 weeks, or 3650 days.

$sel:intervalUnit:CrossRegionCopyDeprecateRule', crossRegionCopyDeprecateRule_intervalUnit - The unit of time in which to measure the Interval.

CrossRegionCopyRetainRule

data CrossRegionCopyRetainRule

Specifies the retention rule for cross-Region snapshot copies.

See: newCrossRegionCopyRetainRule smart constructor.

Constructors

CrossRegionCopyRetainRule' (Maybe Natural) (Maybe RetentionIntervalUnitValues)

newCrossRegionCopyRetainRule :: CrossRegionCopyRetainRule

Create a value of CrossRegionCopyRetainRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:interval:CrossRegionCopyRetainRule', crossRegionCopyRetainRule_interval - The amount of time to retain each snapshot. The maximum is 100 years. This is equivalent to 1200 months, 5200 weeks, or 36500 days.

$sel:intervalUnit:CrossRegionCopyRetainRule', crossRegionCopyRetainRule_intervalUnit - The unit of time for time-based retention.

CrossRegionCopyRule

data CrossRegionCopyRule

Specifies a rule for cross-Region snapshot copies.

See: newCrossRegionCopyRule smart constructor.

Constructors

CrossRegionCopyRule' (Maybe CrossRegionCopyDeprecateRule) (Maybe Text) (Maybe Bool) (Maybe Text) (Maybe CrossRegionCopyRetainRule) (Maybe Text) Bool

newCrossRegionCopyRule

Arguments

:: Bool
$sel:encrypted:CrossRegionCopyRule'

-> CrossRegionCopyRule

Create a value of CrossRegionCopyRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deprecateRule:CrossRegionCopyRule', crossRegionCopyRule_deprecateRule - The AMI deprecation rule for cross-Region AMI copies created by the rule.

$sel:targetRegion:CrossRegionCopyRule', crossRegionCopyRule_targetRegion - Avoid using this parameter when creating new policies. Instead, use Target to specify a target Region or a target Outpost for snapshot copies.

For policies created before the Target parameter was introduced, this parameter indicates the target Region for snapshot copies.

$sel:copyTags:CrossRegionCopyRule', crossRegionCopyRule_copyTags - Indicates whether to copy all user-defined tags from the source snapshot to the cross-Region snapshot copy.

$sel:cmkArn:CrossRegionCopyRule', crossRegionCopyRule_cmkArn - The Amazon Resource Name (ARN) of the KMS key to use for EBS encryption. If this parameter is not specified, the default KMS key for the account is used.

$sel:retainRule:CrossRegionCopyRule', crossRegionCopyRule_retainRule - The retention rule that indicates how long snapshot copies are to be retained in the destination Region.

$sel:target:CrossRegionCopyRule', crossRegionCopyRule_target - The target Region or the Amazon Resource Name (ARN) of the target Outpost for the snapshot copies.

Use this parameter instead of TargetRegion. Do not specify both.

$sel:encrypted:CrossRegionCopyRule', crossRegionCopyRule_encrypted - To encrypt a copy of an unencrypted snapshot if encryption by default is not enabled, enable encryption using this parameter. Copies of encrypted snapshots are encrypted, even if this parameter is false or if encryption by default is not enabled.

DeprecateRule

data DeprecateRule

Specifies an AMI deprecation rule for a schedule.

See: newDeprecateRule smart constructor.

Constructors

DeprecateRule' (Maybe Natural) (Maybe Natural) (Maybe RetentionIntervalUnitValues)

newDeprecateRule :: DeprecateRule

Create a value of DeprecateRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:count:DeprecateRule', deprecateRule_count - If the schedule has a count-based retention rule, this parameter specifies the number of oldest AMIs to deprecate. The count must be less than or equal to the schedule's retention count, and it can't be greater than 1000.

$sel:interval:DeprecateRule', deprecateRule_interval - If the schedule has an age-based retention rule, this parameter specifies the period after which to deprecate AMIs created by the schedule. The period must be less than or equal to the schedule's retention period, and it can't be greater than 10 years. This is equivalent to 120 months, 520 weeks, or 3650 days.

$sel:intervalUnit:DeprecateRule', deprecateRule_intervalUnit - The unit of time in which to measure the Interval.

EncryptionConfiguration

data EncryptionConfiguration

Specifies the encryption settings for shared snapshots that are copied across Regions.

See: newEncryptionConfiguration smart constructor.

Constructors

EncryptionConfiguration' (Maybe Text) Bool

newEncryptionConfiguration

Arguments

:: Bool
$sel:encrypted:EncryptionConfiguration'

-> EncryptionConfiguration

Create a value of EncryptionConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cmkArn:EncryptionConfiguration', encryptionConfiguration_cmkArn - The Amazon Resource Name (ARN) of the KMS key to use for EBS encryption. If this parameter is not specified, the default KMS key for the account is used.

$sel:encrypted:EncryptionConfiguration', encryptionConfiguration_encrypted - To encrypt a copy of an unencrypted snapshot when encryption by default is not enabled, enable encryption using this parameter. Copies of encrypted snapshots are encrypted, even if this parameter is false or when encryption by default is not enabled.

EventParameters

data EventParameters

Specifies an event that triggers an event-based policy.

See: newEventParameters smart constructor.

Constructors

EventParameters' EventTypeValues [Text] Text

newEventParameters

Arguments

:: EventTypeValues
$sel:eventType:EventParameters'

-> Text
$sel:descriptionRegex:EventParameters'

-> EventParameters

Create a value of EventParameters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventType:EventParameters', eventParameters_eventType - The type of event. Currently, only snapshot sharing events are supported.

$sel:snapshotOwner:EventParameters', eventParameters_snapshotOwner - The IDs of the Amazon Web Services accounts that can trigger policy by sharing snapshots with your account. The policy only runs if one of the specified Amazon Web Services accounts shares a snapshot with your account.

$sel:descriptionRegex:EventParameters', eventParameters_descriptionRegex - The snapshot description that can trigger the policy. The description pattern is specified using a regular expression. The policy runs only if a snapshot with a description that matches the specified pattern is shared with your account.

For example, specifying ^.*Created for policy: policy-1234567890abcdef0.*$ configures the policy to run only if snapshots created by policy policy-1234567890abcdef0 are shared with your account.

EventSource

data EventSource

Specifies an event that triggers an event-based policy.

See: newEventSource smart constructor.

Constructors

EventSource' (Maybe EventParameters) EventSourceValues

newEventSource

Arguments

:: EventSourceValues
$sel:type':EventSource'

-> EventSource

Create a value of EventSource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parameters:EventSource', eventSource_parameters - Information about the event.

$sel:type':EventSource', eventSource_type - The source of the event. Currently only managed CloudWatch Events rules are supported.

FastRestoreRule

data FastRestoreRule

Specifies a rule for enabling fast snapshot restore. You can enable fast snapshot restore based on either a count or a time interval.

See: newFastRestoreRule smart constructor.

Constructors

FastRestoreRule' (Maybe Natural) (Maybe Natural) (Maybe RetentionIntervalUnitValues) (NonEmpty Text)

newFastRestoreRule

Arguments

:: NonEmpty Text
$sel:availabilityZones:FastRestoreRule'

-> FastRestoreRule

Create a value of FastRestoreRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:count:FastRestoreRule', fastRestoreRule_count - The number of snapshots to be enabled with fast snapshot restore.

$sel:interval:FastRestoreRule', fastRestoreRule_interval - The amount of time to enable fast snapshot restore. The maximum is 100 years. This is equivalent to 1200 months, 5200 weeks, or 36500 days.

$sel:intervalUnit:FastRestoreRule', fastRestoreRule_intervalUnit - The unit of time for enabling fast snapshot restore.

$sel:availabilityZones:FastRestoreRule', fastRestoreRule_availabilityZones - The Availability Zones in which to enable fast snapshot restore.

LifecyclePolicy

data LifecyclePolicy

Detailed information about a lifecycle policy.

See: newLifecyclePolicy smart constructor.

Constructors

LifecyclePolicy' (Maybe GettablePolicyStateValues) (Maybe PolicyDetails) (Maybe Text) (Maybe Text) (Maybe POSIX) (Maybe Text) (Maybe POSIX) (Maybe Text) (Maybe Text) (Maybe (HashMap Text Text))

newLifecyclePolicy :: LifecyclePolicy

Create a value of LifecyclePolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:LifecyclePolicy', lifecyclePolicy_state - The activation state of the lifecycle policy.

$sel:policyDetails:LifecyclePolicy', lifecyclePolicy_policyDetails - The configuration of the lifecycle policy

$sel:policyId:LifecyclePolicy', lifecyclePolicy_policyId - The identifier of the lifecycle policy.

$sel:executionRoleArn:LifecyclePolicy', lifecyclePolicy_executionRoleArn - The Amazon Resource Name (ARN) of the IAM role used to run the operations specified by the lifecycle policy.

$sel:dateCreated:LifecyclePolicy', lifecyclePolicy_dateCreated - The local date and time when the lifecycle policy was created.

$sel:statusMessage:LifecyclePolicy', lifecyclePolicy_statusMessage - The description of the status.

$sel:dateModified:LifecyclePolicy', lifecyclePolicy_dateModified - The local date and time when the lifecycle policy was last modified.

$sel:policyArn:LifecyclePolicy', lifecyclePolicy_policyArn - The Amazon Resource Name (ARN) of the policy.

$sel:description:LifecyclePolicy', lifecyclePolicy_description - The description of the lifecycle policy.

$sel:tags:LifecyclePolicy', lifecyclePolicy_tags - The tags.

LifecyclePolicySummary

data LifecyclePolicySummary

Summary information about a lifecycle policy.

See: newLifecyclePolicySummary smart constructor.

Constructors

LifecyclePolicySummary' (Maybe GettablePolicyStateValues) (Maybe Text) (Maybe PolicyTypeValues) (Maybe Text) (Maybe (HashMap Text Text))

newLifecyclePolicySummary :: LifecyclePolicySummary

Create a value of LifecyclePolicySummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:LifecyclePolicySummary', lifecyclePolicySummary_state - The activation state of the lifecycle policy.

$sel:policyId:LifecyclePolicySummary', lifecyclePolicySummary_policyId - The identifier of the lifecycle policy.

$sel:policyType:LifecyclePolicySummary', lifecyclePolicySummary_policyType - The type of policy. EBS_SNAPSHOT_MANAGEMENT indicates that the policy manages the lifecycle of Amazon EBS snapshots. IMAGE_MANAGEMENT indicates that the policy manages the lifecycle of EBS-backed AMIs.

$sel:description:LifecyclePolicySummary', lifecyclePolicySummary_description - The description of the lifecycle policy.

$sel:tags:LifecyclePolicySummary', lifecyclePolicySummary_tags - The tags.

Parameters

data Parameters

Specifies optional parameters to add to a policy. The set of valid parameters depends on the combination of policy type and resource type.

See: newParameters smart constructor.

Constructors

Parameters' (Maybe Bool) (Maybe Bool)

newParameters :: Parameters

Create a value of Parameters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:noReboot:Parameters', parameters_noReboot - Applies to AMI lifecycle policies only. Indicates whether targeted instances are rebooted when the lifecycle policy runs. true indicates that targeted instances are not rebooted when the policy runs. false indicates that target instances are rebooted when the policy runs. The default is true (instances are not rebooted).

$sel:excludeBootVolume:Parameters', parameters_excludeBootVolume - [EBS Snapshot Management – Instance policies only] Indicates whether to exclude the root volume from snapshots created using CreateSnapshots. The default is false.

PolicyDetails

data PolicyDetails

Specifies the configuration of a lifecycle policy.

See: newPolicyDetails smart constructor.

Constructors

PolicyDetails' (Maybe (NonEmpty Action)) (Maybe (NonEmpty Tag)) (Maybe PolicyTypeValues) (Maybe (NonEmpty ResourceLocationValues)) (Maybe Parameters) (Maybe (NonEmpty Schedule)) (Maybe EventSource) (Maybe (NonEmpty ResourceTypeValues))

newPolicyDetails :: PolicyDetails

Create a value of PolicyDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:actions:PolicyDetails', policyDetails_actions - The actions to be performed when the event-based policy is triggered. You can specify only one action per policy.

This parameter is required for event-based policies only. If you are creating a snapshot or AMI policy, omit this parameter.

$sel:targetTags:PolicyDetails', policyDetails_targetTags - The single tag that identifies targeted resources for this policy.

This parameter is required for snapshot and AMI policies only. If you are creating an event-based policy, omit this parameter.

$sel:policyType:PolicyDetails', policyDetails_policyType - The valid target resource types and actions a policy can manage. Specify EBS_SNAPSHOT_MANAGEMENT to create a lifecycle policy that manages the lifecycle of Amazon EBS snapshots. Specify IMAGE_MANAGEMENT to create a lifecycle policy that manages the lifecycle of EBS-backed AMIs. Specify EVENT_BASED_POLICY to create an event-based policy that performs specific actions when a defined event occurs in your Amazon Web Services account.

The default is EBS_SNAPSHOT_MANAGEMENT.

$sel:resourceLocations:PolicyDetails', policyDetails_resourceLocations - The location of the resources to backup. If the source resources are located in an Amazon Web Services Region, specify CLOUD. If the source resources are located on an Outpost in your account, specify OUTPOST.

If you specify OUTPOST, Amazon Data Lifecycle Manager backs up all resources of the specified type with matching target tags across all of the Outposts in your account.

$sel:parameters:PolicyDetails', policyDetails_parameters - A set of optional parameters for snapshot and AMI lifecycle policies.

This parameter is required for snapshot and AMI policies only. If you are creating an event-based policy, omit this parameter.

$sel:schedules:PolicyDetails', policyDetails_schedules - The schedules of policy-defined actions for snapshot and AMI lifecycle policies. A policy can have up to four schedules—one mandatory schedule and up to three optional schedules.

This parameter is required for snapshot and AMI policies only. If you are creating an event-based policy, omit this parameter.

$sel:eventSource:PolicyDetails', policyDetails_eventSource - The event that triggers the event-based policy.

This parameter is required for event-based policies only. If you are creating a snapshot or AMI policy, omit this parameter.

$sel:resourceTypes:PolicyDetails', policyDetails_resourceTypes - The target resource type for snapshot and AMI lifecycle policies. Use VOLUME to create snapshots of individual volumes or use INSTANCE to create multi-volume snapshots from the volumes for an instance.

This parameter is required for snapshot and AMI policies only. If you are creating an event-based policy, omit this parameter.

RetainRule

data RetainRule

Specifies the retention rule for a lifecycle policy. You can retain snapshots based on either a count or a time interval.

See: newRetainRule smart constructor.

Constructors

RetainRule' (Maybe Natural) (Maybe Natural) (Maybe RetentionIntervalUnitValues)

newRetainRule :: RetainRule

Create a value of RetainRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:count:RetainRule', retainRule_count - The number of snapshots to retain for each volume, up to a maximum of 1000.

$sel:interval:RetainRule', retainRule_interval - The amount of time to retain each snapshot. The maximum is 100 years. This is equivalent to 1200 months, 5200 weeks, or 36500 days.

$sel:intervalUnit:RetainRule', retainRule_intervalUnit - The unit of time for time-based retention.

Schedule

data Schedule

Specifies a backup schedule for a snapshot or AMI lifecycle policy.

See: newSchedule smart constructor.

Constructors

Schedule' (Maybe [Tag]) (Maybe CreateRule) (Maybe DeprecateRule) (Maybe Bool) (Maybe Text) (Maybe [ShareRule]) (Maybe [Tag]) (Maybe RetainRule) (Maybe [CrossRegionCopyRule]) (Maybe FastRestoreRule)

newSchedule :: Schedule

Create a value of Schedule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:variableTags:Schedule', schedule_variableTags - A collection of key/value pairs with values determined dynamically when the policy is executed. Keys may be any valid Amazon EC2 tag key. Values must be in one of the two following formats: $(instance-id) or $(timestamp). Variable tags are only valid for EBS Snapshot Management – Instance policies.

$sel:createRule:Schedule', schedule_createRule - The creation rule.

$sel:deprecateRule:Schedule', schedule_deprecateRule - The AMI deprecation rule for the schedule.

$sel:copyTags:Schedule', schedule_copyTags - Copy all user-defined tags on a source volume to snapshots of the volume created by this policy.

$sel:name:Schedule', schedule_name - The name of the schedule.

$sel:shareRules:Schedule', schedule_shareRules - The rule for sharing snapshots with other Amazon Web Services accounts.

$sel:tagsToAdd:Schedule', schedule_tagsToAdd - The tags to apply to policy-created resources. These user-defined tags are in addition to the Amazon Web Services-added lifecycle tags.

$sel:retainRule:Schedule', schedule_retainRule - The retention rule.

$sel:crossRegionCopyRules:Schedule', schedule_crossRegionCopyRules - The rule for cross-Region snapshot copies.

You can only specify cross-Region copy rules for policies that create snapshots in a Region. If the policy creates snapshots on an Outpost, then you cannot copy the snapshots to a Region or to an Outpost. If the policy creates snapshots in a Region, then snapshots can be copied to up to three Regions or Outposts.

$sel:fastRestoreRule:Schedule', schedule_fastRestoreRule - The rule for enabling fast snapshot restore.

ShareRule

data ShareRule

Specifies a rule for sharing snapshots across Amazon Web Services accounts.

See: newShareRule smart constructor.

Constructors

ShareRule' (Maybe RetentionIntervalUnitValues) (Maybe Natural) (NonEmpty Text)

newShareRule

Arguments

:: NonEmpty Text
$sel:targetAccounts:ShareRule'

-> ShareRule

Create a value of ShareRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:unshareIntervalUnit:ShareRule', shareRule_unshareIntervalUnit - The unit of time for the automatic unsharing interval.

$sel:unshareInterval:ShareRule', shareRule_unshareInterval - The period after which snapshots that are shared with other Amazon Web Services accounts are automatically unshared.

$sel:targetAccounts:ShareRule', shareRule_targetAccounts - The IDs of the Amazon Web Services accounts with which to share the snapshots.

Tag

data Tag

Specifies a tag for a resource.

See: newTag smart constructor.

Constructors

Tag' Text Text

newTag

Arguments

:: Text
$sel:key:Tag'

-> Text
$sel:value:Tag'

-> Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:Tag', tag_key - The tag key.

$sel:value:Tag', tag_value - The tag value.