Network.AWS.DLM.Lens

Description

Operations

DeleteLifecyclePolicy

deleteLifecyclePolicy_policyId :: Lens' DeleteLifecyclePolicy Text

The identifier of the lifecycle policy.

deleteLifecyclePolicyResponse_httpStatus :: Lens' DeleteLifecyclePolicyResponse Int

The response's http status code.

UpdateLifecyclePolicy

updateLifecyclePolicy_state :: Lens' UpdateLifecyclePolicy (Maybe SettablePolicyStateValues)

The desired activation state of the lifecycle policy after creation.

updateLifecyclePolicy_policyDetails :: Lens' UpdateLifecyclePolicy (Maybe PolicyDetails)

The configuration of the lifecycle policy. You cannot update the policy type or the resource type.

updateLifecyclePolicy_executionRoleArn :: Lens' UpdateLifecyclePolicy (Maybe Text)

The Amazon Resource Name (ARN) of the IAM role used to run the operations specified by the lifecycle policy.

updateLifecyclePolicy_description :: Lens' UpdateLifecyclePolicy (Maybe Text)

A description of the lifecycle policy.

updateLifecyclePolicy_policyId :: Lens' UpdateLifecyclePolicy Text

The identifier of the lifecycle policy.

updateLifecyclePolicyResponse_httpStatus :: Lens' UpdateLifecyclePolicyResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The Amazon Resource Name (ARN) of the resource.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe (HashMap Text Text))

Information about the tags.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

CreateLifecyclePolicy

createLifecyclePolicy_tags :: Lens' CreateLifecyclePolicy (Maybe (HashMap Text Text))

The tags to apply to the lifecycle policy during creation.

createLifecyclePolicy_executionRoleArn :: Lens' CreateLifecyclePolicy Text

The Amazon Resource Name (ARN) of the IAM role used to run the operations specified by the lifecycle policy.

createLifecyclePolicy_description :: Lens' CreateLifecyclePolicy Text

A description of the lifecycle policy. The characters ^[0-9A-Za-z _-]+$ are supported.

createLifecyclePolicy_state :: Lens' CreateLifecyclePolicy SettablePolicyStateValues

The desired activation state of the lifecycle policy after creation.

createLifecyclePolicy_policyDetails :: Lens' CreateLifecyclePolicy PolicyDetails

The configuration details of the lifecycle policy.

createLifecyclePolicyResponse_policyId :: Lens' CreateLifecyclePolicyResponse (Maybe Text)

The identifier of the lifecycle policy.

createLifecyclePolicyResponse_httpStatus :: Lens' CreateLifecyclePolicyResponse Int

The response's http status code.

GetLifecyclePolicy

getLifecyclePolicy_policyId :: Lens' GetLifecyclePolicy Text

The identifier of the lifecycle policy.

getLifecyclePolicyResponse_policy :: Lens' GetLifecyclePolicyResponse (Maybe LifecyclePolicy)

Detailed information about the lifecycle policy.

getLifecyclePolicyResponse_httpStatus :: Lens' GetLifecyclePolicyResponse Int

The response's http status code.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The Amazon Resource Name (ARN) of the resource.

tagResource_tags :: Lens' TagResource (HashMap Text Text)

One or more tags.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

UntagResource

untagResource_resourceArn :: Lens' UntagResource Text

The Amazon Resource Name (ARN) of the resource.

untagResource_tagKeys :: Lens' UntagResource (NonEmpty Text)

The tag keys.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

GetLifecyclePolicies

getLifecyclePolicies_state :: Lens' GetLifecyclePolicies (Maybe GettablePolicyStateValues)

The activation state.

getLifecyclePolicies_targetTags :: Lens' GetLifecyclePolicies (Maybe (NonEmpty Text))

The target tag for a policy.

Tags are strings in the format key=value.

getLifecyclePolicies_tagsToAdd :: Lens' GetLifecyclePolicies (Maybe [Text])

The tags to add to objects created by the policy.

Tags are strings in the format key=value.

These user-defined tags are added in addition to the Amazon Web Services-added lifecycle tags.

getLifecyclePolicies_policyIds :: Lens' GetLifecyclePolicies (Maybe [Text])

The identifiers of the data lifecycle policies.

getLifecyclePolicies_resourceTypes :: Lens' GetLifecyclePolicies (Maybe (NonEmpty ResourceTypeValues))

The resource type.

getLifecyclePoliciesResponse_policies :: Lens' GetLifecyclePoliciesResponse (Maybe [LifecyclePolicySummary])

Summary information about the lifecycle policies.

getLifecyclePoliciesResponse_httpStatus :: Lens' GetLifecyclePoliciesResponse Int

The response's http status code.

Types

Action

action_name :: Lens' Action Text

A descriptive name for the action.

action_crossRegionCopy :: Lens' Action [CrossRegionCopyAction]

The rule for copying shared snapshots across Regions.

CreateRule

createRule_location :: Lens' CreateRule (Maybe LocationValues)

Specifies the destination for snapshots created by the policy. To create snapshots in the same Region as the source resource, specify CLOUD. To create snapshots on the same Outpost as the source resource, specify OUTPOST_LOCAL. If you omit this parameter, CLOUD is used by default.

If the policy targets resources in an Amazon Web Services Region, then you must create snapshots in the same Region as the source resource.

If the policy targets resources on an Outpost, then you can create snapshots on the same Outpost as the source resource, or in the Region of that Outpost.

createRule_interval :: Lens' CreateRule (Maybe Natural)

The interval between snapshots. The supported values are 1, 2, 3, 4, 6, 8, 12, and 24.

createRule_cronExpression :: Lens' CreateRule (Maybe Text)

The schedule, as a Cron expression. The schedule interval must be between 1 hour and 1 year. For more information, see Cron expressions in the Amazon CloudWatch User Guide.

createRule_times :: Lens' CreateRule (Maybe [Text])

The time, in UTC, to start the operation. The supported format is hh:mm.

The operation occurs within a one-hour window following the specified time. If you do not specify a time, Amazon DLM selects a time within the next 24 hours.

createRule_intervalUnit :: Lens' CreateRule (Maybe IntervalUnitValues)

The interval unit.

CrossRegionCopyAction

crossRegionCopyAction_retainRule :: Lens' CrossRegionCopyAction (Maybe CrossRegionCopyRetainRule)

Undocumented member.

crossRegionCopyAction_target :: Lens' CrossRegionCopyAction Text

The target Region.

crossRegionCopyAction_encryptionConfiguration :: Lens' CrossRegionCopyAction EncryptionConfiguration

The encryption settings for the copied snapshot.

CrossRegionCopyDeprecateRule

crossRegionCopyDeprecateRule_interval :: Lens' CrossRegionCopyDeprecateRule (Maybe Natural)

The period after which to deprecate the cross-Region AMI copies. The period must be less than or equal to the cross-Region AMI copy retention period, and it can't be greater than 10 years. This is equivalent to 120 months, 520 weeks, or 3650 days.

crossRegionCopyDeprecateRule_intervalUnit :: Lens' CrossRegionCopyDeprecateRule (Maybe RetentionIntervalUnitValues)

The unit of time in which to measure the Interval.

CrossRegionCopyRetainRule

crossRegionCopyRetainRule_interval :: Lens' CrossRegionCopyRetainRule (Maybe Natural)

The amount of time to retain each snapshot. The maximum is 100 years. This is equivalent to 1200 months, 5200 weeks, or 36500 days.

crossRegionCopyRetainRule_intervalUnit :: Lens' CrossRegionCopyRetainRule (Maybe RetentionIntervalUnitValues)

The unit of time for time-based retention.

CrossRegionCopyRule

crossRegionCopyRule_deprecateRule :: Lens' CrossRegionCopyRule (Maybe CrossRegionCopyDeprecateRule)

The AMI deprecation rule for cross-Region AMI copies created by the rule.

crossRegionCopyRule_targetRegion :: Lens' CrossRegionCopyRule (Maybe Text)

Avoid using this parameter when creating new policies. Instead, use Target to specify a target Region or a target Outpost for snapshot copies.

For policies created before the Target parameter was introduced, this parameter indicates the target Region for snapshot copies.

crossRegionCopyRule_copyTags :: Lens' CrossRegionCopyRule (Maybe Bool)

Indicates whether to copy all user-defined tags from the source snapshot to the cross-Region snapshot copy.

crossRegionCopyRule_cmkArn :: Lens' CrossRegionCopyRule (Maybe Text)

The Amazon Resource Name (ARN) of the KMS key to use for EBS encryption. If this parameter is not specified, the default KMS key for the account is used.

crossRegionCopyRule_retainRule :: Lens' CrossRegionCopyRule (Maybe CrossRegionCopyRetainRule)

The retention rule that indicates how long snapshot copies are to be retained in the destination Region.

crossRegionCopyRule_target :: Lens' CrossRegionCopyRule (Maybe Text)

The target Region or the Amazon Resource Name (ARN) of the target Outpost for the snapshot copies.

Use this parameter instead of TargetRegion. Do not specify both.

crossRegionCopyRule_encrypted :: Lens' CrossRegionCopyRule Bool

To encrypt a copy of an unencrypted snapshot if encryption by default is not enabled, enable encryption using this parameter. Copies of encrypted snapshots are encrypted, even if this parameter is false or if encryption by default is not enabled.

DeprecateRule

deprecateRule_count :: Lens' DeprecateRule (Maybe Natural)

If the schedule has a count-based retention rule, this parameter specifies the number of oldest AMIs to deprecate. The count must be less than or equal to the schedule's retention count, and it can't be greater than 1000.

deprecateRule_interval :: Lens' DeprecateRule (Maybe Natural)

If the schedule has an age-based retention rule, this parameter specifies the period after which to deprecate AMIs created by the schedule. The period must be less than or equal to the schedule's retention period, and it can't be greater than 10 years. This is equivalent to 120 months, 520 weeks, or 3650 days.

deprecateRule_intervalUnit :: Lens' DeprecateRule (Maybe RetentionIntervalUnitValues)

The unit of time in which to measure the Interval.

EncryptionConfiguration

encryptionConfiguration_cmkArn :: Lens' EncryptionConfiguration (Maybe Text)

The Amazon Resource Name (ARN) of the KMS key to use for EBS encryption. If this parameter is not specified, the default KMS key for the account is used.

encryptionConfiguration_encrypted :: Lens' EncryptionConfiguration Bool

To encrypt a copy of an unencrypted snapshot when encryption by default is not enabled, enable encryption using this parameter. Copies of encrypted snapshots are encrypted, even if this parameter is false or when encryption by default is not enabled.

EventParameters

eventParameters_eventType :: Lens' EventParameters EventTypeValues

The type of event. Currently, only snapshot sharing events are supported.

eventParameters_snapshotOwner :: Lens' EventParameters [Text]

The IDs of the Amazon Web Services accounts that can trigger policy by sharing snapshots with your account. The policy only runs if one of the specified Amazon Web Services accounts shares a snapshot with your account.

eventParameters_descriptionRegex :: Lens' EventParameters Text

The snapshot description that can trigger the policy. The description pattern is specified using a regular expression. The policy runs only if a snapshot with a description that matches the specified pattern is shared with your account.

For example, specifying ^.*Created for policy: policy-1234567890abcdef0.*$ configures the policy to run only if snapshots created by policy policy-1234567890abcdef0 are shared with your account.

EventSource

eventSource_parameters :: Lens' EventSource (Maybe EventParameters)

Information about the event.

eventSource_type :: Lens' EventSource EventSourceValues

The source of the event. Currently only managed CloudWatch Events rules are supported.

FastRestoreRule

fastRestoreRule_count :: Lens' FastRestoreRule (Maybe Natural)

The number of snapshots to be enabled with fast snapshot restore.

fastRestoreRule_interval :: Lens' FastRestoreRule (Maybe Natural)

The amount of time to enable fast snapshot restore. The maximum is 100 years. This is equivalent to 1200 months, 5200 weeks, or 36500 days.

fastRestoreRule_intervalUnit :: Lens' FastRestoreRule (Maybe RetentionIntervalUnitValues)

The unit of time for enabling fast snapshot restore.

fastRestoreRule_availabilityZones :: Lens' FastRestoreRule (NonEmpty Text)

The Availability Zones in which to enable fast snapshot restore.

LifecyclePolicy

lifecyclePolicy_state :: Lens' LifecyclePolicy (Maybe GettablePolicyStateValues)

The activation state of the lifecycle policy.

lifecyclePolicy_policyDetails :: Lens' LifecyclePolicy (Maybe PolicyDetails)

The configuration of the lifecycle policy

lifecyclePolicy_policyId :: Lens' LifecyclePolicy (Maybe Text)

The identifier of the lifecycle policy.

lifecyclePolicy_executionRoleArn :: Lens' LifecyclePolicy (Maybe Text)

The Amazon Resource Name (ARN) of the IAM role used to run the operations specified by the lifecycle policy.

lifecyclePolicy_dateCreated :: Lens' LifecyclePolicy (Maybe UTCTime)

The local date and time when the lifecycle policy was created.

lifecyclePolicy_statusMessage :: Lens' LifecyclePolicy (Maybe Text)

The description of the status.

lifecyclePolicy_dateModified :: Lens' LifecyclePolicy (Maybe UTCTime)

The local date and time when the lifecycle policy was last modified.

lifecyclePolicy_policyArn :: Lens' LifecyclePolicy (Maybe Text)

The Amazon Resource Name (ARN) of the policy.

lifecyclePolicy_description :: Lens' LifecyclePolicy (Maybe Text)

The description of the lifecycle policy.

lifecyclePolicy_tags :: Lens' LifecyclePolicy (Maybe (HashMap Text Text))

The tags.

LifecyclePolicySummary

lifecyclePolicySummary_state :: Lens' LifecyclePolicySummary (Maybe GettablePolicyStateValues)

The activation state of the lifecycle policy.

lifecyclePolicySummary_policyId :: Lens' LifecyclePolicySummary (Maybe Text)

The identifier of the lifecycle policy.

lifecyclePolicySummary_policyType :: Lens' LifecyclePolicySummary (Maybe PolicyTypeValues)

The type of policy. EBS_SNAPSHOT_MANAGEMENT indicates that the policy manages the lifecycle of Amazon EBS snapshots. IMAGE_MANAGEMENT indicates that the policy manages the lifecycle of EBS-backed AMIs.

lifecyclePolicySummary_description :: Lens' LifecyclePolicySummary (Maybe Text)

The description of the lifecycle policy.

lifecyclePolicySummary_tags :: Lens' LifecyclePolicySummary (Maybe (HashMap Text Text))

The tags.

Parameters

parameters_noReboot :: Lens' Parameters (Maybe Bool)

Applies to AMI lifecycle policies only. Indicates whether targeted instances are rebooted when the lifecycle policy runs. true indicates that targeted instances are not rebooted when the policy runs. false indicates that target instances are rebooted when the policy runs. The default is true (instances are not rebooted).

parameters_excludeBootVolume :: Lens' Parameters (Maybe Bool)

EBS Snapshot Management – Instance policies only

Indicates whether to exclude the root volume from snapshots created using CreateSnapshots. The default is false.

PolicyDetails

policyDetails_actions :: Lens' PolicyDetails (Maybe (NonEmpty Action))

The actions to be performed when the event-based policy is triggered. You can specify only one action per policy.

This parameter is required for event-based policies only. If you are creating a snapshot or AMI policy, omit this parameter.

policyDetails_targetTags :: Lens' PolicyDetails (Maybe (NonEmpty Tag))

The single tag that identifies targeted resources for this policy.

This parameter is required for snapshot and AMI policies only. If you are creating an event-based policy, omit this parameter.

policyDetails_policyType :: Lens' PolicyDetails (Maybe PolicyTypeValues)

The valid target resource types and actions a policy can manage. Specify EBS_SNAPSHOT_MANAGEMENT to create a lifecycle policy that manages the lifecycle of Amazon EBS snapshots. Specify IMAGE_MANAGEMENT to create a lifecycle policy that manages the lifecycle of EBS-backed AMIs. Specify EVENT_BASED_POLICY to create an event-based policy that performs specific actions when a defined event occurs in your Amazon Web Services account.

The default is EBS_SNAPSHOT_MANAGEMENT.

policyDetails_resourceLocations :: Lens' PolicyDetails (Maybe (NonEmpty ResourceLocationValues))

The location of the resources to backup. If the source resources are located in an Amazon Web Services Region, specify CLOUD. If the source resources are located on an Outpost in your account, specify OUTPOST.

If you specify OUTPOST, Amazon Data Lifecycle Manager backs up all resources of the specified type with matching target tags across all of the Outposts in your account.

policyDetails_parameters :: Lens' PolicyDetails (Maybe Parameters)

A set of optional parameters for snapshot and AMI lifecycle policies.

This parameter is required for snapshot and AMI policies only. If you are creating an event-based policy, omit this parameter.

policyDetails_schedules :: Lens' PolicyDetails (Maybe (NonEmpty Schedule))

The schedules of policy-defined actions for snapshot and AMI lifecycle policies. A policy can have up to four schedules—one mandatory schedule and up to three optional schedules.

This parameter is required for snapshot and AMI policies only. If you are creating an event-based policy, omit this parameter.

policyDetails_eventSource :: Lens' PolicyDetails (Maybe EventSource)

The event that triggers the event-based policy.

This parameter is required for event-based policies only. If you are creating a snapshot or AMI policy, omit this parameter.

policyDetails_resourceTypes :: Lens' PolicyDetails (Maybe (NonEmpty ResourceTypeValues))

The target resource type for snapshot and AMI lifecycle policies. Use VOLUME to create snapshots of individual volumes or use INSTANCE to create multi-volume snapshots from the volumes for an instance.

This parameter is required for snapshot and AMI policies only. If you are creating an event-based policy, omit this parameter.

RetainRule

retainRule_count :: Lens' RetainRule (Maybe Natural)

The number of snapshots to retain for each volume, up to a maximum of 1000.

retainRule_interval :: Lens' RetainRule (Maybe Natural)

The amount of time to retain each snapshot. The maximum is 100 years. This is equivalent to 1200 months, 5200 weeks, or 36500 days.

retainRule_intervalUnit :: Lens' RetainRule (Maybe RetentionIntervalUnitValues)

The unit of time for time-based retention.

Schedule

schedule_variableTags :: Lens' Schedule (Maybe [Tag])

A collection of key/value pairs with values determined dynamically when the policy is executed. Keys may be any valid Amazon EC2 tag key. Values must be in one of the two following formats: $(instance-id) or $(timestamp). Variable tags are only valid for EBS Snapshot Management – Instance policies.

schedule_createRule :: Lens' Schedule (Maybe CreateRule)

The creation rule.

schedule_deprecateRule :: Lens' Schedule (Maybe DeprecateRule)

The AMI deprecation rule for the schedule.

schedule_copyTags :: Lens' Schedule (Maybe Bool)

Copy all user-defined tags on a source volume to snapshots of the volume created by this policy.

schedule_name :: Lens' Schedule (Maybe Text)

The name of the schedule.

schedule_shareRules :: Lens' Schedule (Maybe [ShareRule])

The rule for sharing snapshots with other Amazon Web Services accounts.

schedule_tagsToAdd :: Lens' Schedule (Maybe [Tag])

The tags to apply to policy-created resources. These user-defined tags are in addition to the Amazon Web Services-added lifecycle tags.

schedule_retainRule :: Lens' Schedule (Maybe RetainRule)

The retention rule.

schedule_crossRegionCopyRules :: Lens' Schedule (Maybe [CrossRegionCopyRule])

The rule for cross-Region snapshot copies.

You can only specify cross-Region copy rules for policies that create snapshots in a Region. If the policy creates snapshots on an Outpost, then you cannot copy the snapshots to a Region or to an Outpost. If the policy creates snapshots in a Region, then snapshots can be copied to up to three Regions or Outposts.

schedule_fastRestoreRule :: Lens' Schedule (Maybe FastRestoreRule)

The rule for enabling fast snapshot restore.

ShareRule

shareRule_unshareIntervalUnit :: Lens' ShareRule (Maybe RetentionIntervalUnitValues)

The unit of time for the automatic unsharing interval.

shareRule_unshareInterval :: Lens' ShareRule (Maybe Natural)

The period after which snapshots that are shared with other Amazon Web Services accounts are automatically unshared.

shareRule_targetAccounts :: Lens' ShareRule (NonEmpty Text)

The IDs of the Amazon Web Services accounts with which to share the snapshots.

Tag

tag_key :: Lens' Tag Text

The tag key.

tag_value :: Lens' Tag Text

The tag value.