Network.AWS.DLM.Types.LifecyclePolicySummary

Description

Documentation

data LifecyclePolicySummary

Summary information about a lifecycle policy.

See: newLifecyclePolicySummary smart constructor.

Constructors

LifecyclePolicySummary'

Fields

state :: Maybe GettablePolicyStateValues

The activation state of the lifecycle policy.

policyId :: Maybe Text

The identifier of the lifecycle policy.

policyType :: Maybe PolicyTypeValues

The type of policy. EBS_SNAPSHOT_MANAGEMENT indicates that the policy manages the lifecycle of Amazon EBS snapshots. IMAGE_MANAGEMENT indicates that the policy manages the lifecycle of EBS-backed AMIs.

description :: Maybe Text

The description of the lifecycle policy.

tags :: Maybe (HashMap Text Text)

The tags.

newLifecyclePolicySummary :: LifecyclePolicySummary

Create a value of LifecyclePolicySummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:LifecyclePolicySummary', lifecyclePolicySummary_state - The activation state of the lifecycle policy.

$sel:policyId:LifecyclePolicySummary', lifecyclePolicySummary_policyId - The identifier of the lifecycle policy.

$sel:policyType:LifecyclePolicySummary', lifecyclePolicySummary_policyType - The type of policy. EBS_SNAPSHOT_MANAGEMENT indicates that the policy manages the lifecycle of Amazon EBS snapshots. IMAGE_MANAGEMENT indicates that the policy manages the lifecycle of EBS-backed AMIs.

$sel:description:LifecyclePolicySummary', lifecyclePolicySummary_description - The description of the lifecycle policy.

$sel:tags:LifecyclePolicySummary', lifecyclePolicySummary_tags - The tags.

lifecyclePolicySummary_state :: Lens' LifecyclePolicySummary (Maybe GettablePolicyStateValues)

The activation state of the lifecycle policy.

lifecyclePolicySummary_policyId :: Lens' LifecyclePolicySummary (Maybe Text)

The identifier of the lifecycle policy.

lifecyclePolicySummary_policyType :: Lens' LifecyclePolicySummary (Maybe PolicyTypeValues)

The type of policy. EBS_SNAPSHOT_MANAGEMENT indicates that the policy manages the lifecycle of Amazon EBS snapshots. IMAGE_MANAGEMENT indicates that the policy manages the lifecycle of EBS-backed AMIs.

lifecyclePolicySummary_description :: Lens' LifecyclePolicySummary (Maybe Text)

The description of the lifecycle policy.

lifecyclePolicySummary_tags :: Lens' LifecyclePolicySummary (Maybe (HashMap Text Text))

The tags.