Network.AWS.DMS.CreateEndpoint

Description

Creates an endpoint using the provided settings.

For a MySQL source or target endpoint, don't explicitly specify the database using the DatabaseName request parameter on the CreateEndpoint API call. Specifying DatabaseName when you create a MySQL endpoint replicates all the task tables to this single database. For MySQL endpoints, you specify the database only when you specify the schema in the table-mapping rules of the DMS task.

Creating a Request

data CreateEndpoint

See: newCreateEndpoint smart constructor.

Constructors

CreateEndpoint'

Fields

mongoDbSettings :: Maybe MongoDbSettings

Settings in JSON format for the source MongoDB endpoint. For more information about the available settings, see Endpoint configuration settings when using MongoDB as a source for Database Migration Service in the Database Migration Service User Guide.

sslMode :: Maybe DmsSslModeValue

The Secure Sockets Layer (SSL) mode to use for the SSL connection. The default is none

neptuneSettings :: Maybe NeptuneSettings

Settings in JSON format for the target Amazon Neptune endpoint. For more information about the available settings, see Specifying graph-mapping rules using Gremlin and R2RML for Amazon Neptune as a target in the Database Migration Service User Guide.

redisSettings :: Maybe RedisSettings

Settings in JSON format for the target Redis endpoint.

externalTableDefinition :: Maybe Text

The external table definition.

elasticsearchSettings :: Maybe ElasticsearchSettings

Settings in JSON format for the target Elasticsearch endpoint. For more information about the available settings, see Extra Connection Attributes When Using Elasticsearch as a Target for DMS in the Database Migration Service User Guide.

postgreSQLSettings :: Maybe PostgreSQLSettings

Settings in JSON format for the source and target PostgreSQL endpoint. For information about other available settings, see Extra connection attributes when using PostgreSQL as a source for DMS and Extra connection attributes when using PostgreSQL as a target for DMS in the Database Migration Service User Guide.

oracleSettings :: Maybe OracleSettings

Settings in JSON format for the source and target Oracle endpoint. For information about other available settings, see Extra connection attributes when using Oracle as a source for DMS and Extra connection attributes when using Oracle as a target for DMS in the Database Migration Service User Guide.

certificateArn :: Maybe Text

The Amazon Resource Name (ARN) for the certificate.

serviceAccessRoleArn :: Maybe Text

The Amazon Resource Name (ARN) for the service access role that you want to use to create the endpoint. The role must allow the iam:PassRole action.

s3Settings :: Maybe S3Settings

Settings in JSON format for the target Amazon S3 endpoint. For more information about the available settings, see Extra Connection Attributes When Using Amazon S3 as a Target for DMS in the Database Migration Service User Guide.

microsoftSQLServerSettings :: Maybe MicrosoftSQLServerSettings

Settings in JSON format for the source and target Microsoft SQL Server endpoint. For information about other available settings, see Extra connection attributes when using SQL Server as a source for DMS and Extra connection attributes when using SQL Server as a target for DMS in the Database Migration Service User Guide.

serverName :: Maybe Text

The name of the server where the endpoint database resides.

kmsKeyId :: Maybe Text

An KMS key identifier that is used to encrypt the connection parameters for the endpoint.

If you don't specify a value for the KmsKeyId parameter, then DMS uses your default encryption key.

KMS creates the default encryption key for your Amazon Web Services account. Your Amazon Web Services account has a different default encryption key for each Amazon Web Services Region.

iBMDb2Settings :: Maybe IBMDb2Settings

Settings in JSON format for the source IBM Db2 LUW endpoint. For information about other available settings, see Extra connection attributes when using Db2 LUW as a source for DMS in the Database Migration Service User Guide.

password :: Maybe (Sensitive Text)

The password to be used to log in to the endpoint database.

mySQLSettings :: Maybe MySQLSettings

Settings in JSON format for the source and target MySQL endpoint. For information about other available settings, see Extra connection attributes when using MySQL as a source for DMS and Extra connection attributes when using a MySQL-compatible database as a target for DMS in the Database Migration Service User Guide.

dmsTransferSettings :: Maybe DmsTransferSettings

The settings in JSON format for the DMS transfer type of source endpoint.

Possible settings include the following:

ServiceAccessRoleArn - The Amazon Resource Name (ARN) used by the service access IAM role. The role must allow the iam:PassRole action.

BucketName - The name of the S3 bucket to use.

Shorthand syntax for these settings is as follows: ServiceAccessRoleArn=string,BucketName=string

JSON syntax for these settings is as follows: { "ServiceAccessRoleArn": "string", "BucketName": "string", }

tags :: Maybe [Tag]

One or more tags to be assigned to the endpoint.

resourceIdentifier :: Maybe Text

A friendly name for the resource identifier at the end of the EndpointArn response parameter that is returned in the created Endpoint object. The value for this parameter can have up to 31 characters. It can contain only ASCII letters, digits, and hyphen ('-'). Also, it can't end with a hyphen or contain two consecutive hyphens, and can only begin with a letter, such as Example-App-ARN1. For example, this value might result in the EndpointArn value arn:aws:dms:eu-west-1:012345678901:rep:Example-App-ARN1. If you don't specify a ResourceIdentifier value, DMS generates a default identifier value for the end of EndpointArn.

port :: Maybe Int

The port used by the endpoint database.

redshiftSettings :: Maybe RedshiftSettings

username :: Maybe Text

The user name to be used to log in to the endpoint database.

extraConnectionAttributes :: Maybe Text

Additional attributes associated with the connection. Each attribute is specified as a name-value pair associated by an equal sign (=). Multiple attributes are separated by a semicolon (;) with no additional white space. For information on the attributes available for connecting your source or target endpoint, see Working with DMS Endpoints in the Database Migration Service User Guide.

kafkaSettings :: Maybe KafkaSettings

Settings in JSON format for the target Apache Kafka endpoint. For more information about the available settings, see Using object mapping to migrate data to a Kafka topic in the Database Migration Service User Guide.

docDbSettings :: Maybe DocDbSettings

dynamoDbSettings :: Maybe DynamoDbSettings

Settings in JSON format for the target Amazon DynamoDB endpoint. For information about other available settings, see Using Object Mapping to Migrate Data to DynamoDB in the Database Migration Service User Guide.

kinesisSettings :: Maybe KinesisSettings

Settings in JSON format for the target endpoint for Amazon Kinesis Data Streams. For more information about the available settings, see Using object mapping to migrate data to a Kinesis data stream in the Database Migration Service User Guide.

sybaseSettings :: Maybe SybaseSettings

Settings in JSON format for the source and target SAP ASE endpoint. For information about other available settings, see Extra connection attributes when using SAP ASE as a source for DMS and Extra connection attributes when using SAP ASE as a target for DMS in the Database Migration Service User Guide.

databaseName :: Maybe Text

The name of the endpoint database. For a MySQL source or target endpoint, do not specify DatabaseName.

endpointIdentifier :: Text

The database endpoint identifier. Identifiers must begin with a letter and must contain only ASCII letters, digits, and hyphens. They can't end with a hyphen, or contain two consecutive hyphens.

endpointType :: ReplicationEndpointTypeValue

The type of endpoint. Valid values are source and target.

engineName :: Text

The type of engine for the endpoint. Valid values, depending on the EndpointType value, include "mysql", "oracle", "postgres", "mariadb", "aurora", "aurora-postgresql", "redshift", "s3", "db2", "azuredb", "sybase", "dynamodb", "mongodb", "kinesis", "kafka", "elasticsearch", "docdb", "sqlserver", and "neptune".

newCreateEndpoint

Arguments

:: Text
$sel:endpointIdentifier:CreateEndpoint'

-> ReplicationEndpointTypeValue
$sel:endpointType:CreateEndpoint'

-> Text
$sel:engineName:CreateEndpoint'

-> CreateEndpoint

Create a value of CreateEndpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:mongoDbSettings:CreateEndpoint', createEndpoint_mongoDbSettings - Settings in JSON format for the source MongoDB endpoint. For more information about the available settings, see Endpoint configuration settings when using MongoDB as a source for Database Migration Service in the Database Migration Service User Guide.

$sel:sslMode:CreateEndpoint', createEndpoint_sslMode - The Secure Sockets Layer (SSL) mode to use for the SSL connection. The default is none

$sel:neptuneSettings:CreateEndpoint', createEndpoint_neptuneSettings - Settings in JSON format for the target Amazon Neptune endpoint. For more information about the available settings, see Specifying graph-mapping rules using Gremlin and R2RML for Amazon Neptune as a target in the Database Migration Service User Guide.

$sel:redisSettings:CreateEndpoint', createEndpoint_redisSettings - Settings in JSON format for the target Redis endpoint.

$sel:externalTableDefinition:CreateEndpoint', createEndpoint_externalTableDefinition - The external table definition.

$sel:elasticsearchSettings:CreateEndpoint', createEndpoint_elasticsearchSettings - Settings in JSON format for the target Elasticsearch endpoint. For more information about the available settings, see Extra Connection Attributes When Using Elasticsearch as a Target for DMS in the Database Migration Service User Guide.

$sel:postgreSQLSettings:CreateEndpoint', createEndpoint_postgreSQLSettings - Settings in JSON format for the source and target PostgreSQL endpoint. For information about other available settings, see Extra connection attributes when using PostgreSQL as a source for DMS and Extra connection attributes when using PostgreSQL as a target for DMS in the Database Migration Service User Guide.

$sel:oracleSettings:CreateEndpoint', createEndpoint_oracleSettings - Settings in JSON format for the source and target Oracle endpoint. For information about other available settings, see Extra connection attributes when using Oracle as a source for DMS and Extra connection attributes when using Oracle as a target for DMS in the Database Migration Service User Guide.

$sel:certificateArn:CreateEndpoint', createEndpoint_certificateArn - The Amazon Resource Name (ARN) for the certificate.

$sel:serviceAccessRoleArn:CreateEndpoint', createEndpoint_serviceAccessRoleArn - The Amazon Resource Name (ARN) for the service access role that you want to use to create the endpoint. The role must allow the iam:PassRole action.

$sel:s3Settings:CreateEndpoint', createEndpoint_s3Settings - Settings in JSON format for the target Amazon S3 endpoint. For more information about the available settings, see Extra Connection Attributes When Using Amazon S3 as a Target for DMS in the Database Migration Service User Guide.

$sel:microsoftSQLServerSettings:CreateEndpoint', createEndpoint_microsoftSQLServerSettings - Settings in JSON format for the source and target Microsoft SQL Server endpoint. For information about other available settings, see Extra connection attributes when using SQL Server as a source for DMS and Extra connection attributes when using SQL Server as a target for DMS in the Database Migration Service User Guide.

$sel:serverName:CreateEndpoint', createEndpoint_serverName - The name of the server where the endpoint database resides.

$sel:kmsKeyId:CreateEndpoint', createEndpoint_kmsKeyId - An KMS key identifier that is used to encrypt the connection parameters for the endpoint.

If you don't specify a value for the KmsKeyId parameter, then DMS uses your default encryption key.

KMS creates the default encryption key for your Amazon Web Services account. Your Amazon Web Services account has a different default encryption key for each Amazon Web Services Region.

$sel:iBMDb2Settings:CreateEndpoint', createEndpoint_iBMDb2Settings - Settings in JSON format for the source IBM Db2 LUW endpoint. For information about other available settings, see Extra connection attributes when using Db2 LUW as a source for DMS in the Database Migration Service User Guide.

$sel:password:CreateEndpoint', createEndpoint_password - The password to be used to log in to the endpoint database.

$sel:mySQLSettings:CreateEndpoint', createEndpoint_mySQLSettings - Settings in JSON format for the source and target MySQL endpoint. For information about other available settings, see Extra connection attributes when using MySQL as a source for DMS and Extra connection attributes when using a MySQL-compatible database as a target for DMS in the Database Migration Service User Guide.

$sel:dmsTransferSettings:CreateEndpoint', createEndpoint_dmsTransferSettings - The settings in JSON format for the DMS transfer type of source endpoint.

Possible settings include the following:

ServiceAccessRoleArn - The Amazon Resource Name (ARN) used by the service access IAM role. The role must allow the iam:PassRole action.

BucketName - The name of the S3 bucket to use.

Shorthand syntax for these settings is as follows: ServiceAccessRoleArn=string,BucketName=string

JSON syntax for these settings is as follows: { "ServiceAccessRoleArn": "string", "BucketName": "string", }

$sel:tags:CreateEndpoint', createEndpoint_tags - One or more tags to be assigned to the endpoint.

$sel:resourceIdentifier:CreateEndpoint', createEndpoint_resourceIdentifier - A friendly name for the resource identifier at the end of the EndpointArn response parameter that is returned in the created Endpoint object. The value for this parameter can have up to 31 characters. It can contain only ASCII letters, digits, and hyphen ('-'). Also, it can't end with a hyphen or contain two consecutive hyphens, and can only begin with a letter, such as Example-App-ARN1. For example, this value might result in the EndpointArn value arn:aws:dms:eu-west-1:012345678901:rep:Example-App-ARN1. If you don't specify a ResourceIdentifier value, DMS generates a default identifier value for the end of EndpointArn.

$sel:port:CreateEndpoint', createEndpoint_port - The port used by the endpoint database.

$sel:redshiftSettings:CreateEndpoint', createEndpoint_redshiftSettings - Undocumented member.

$sel:username:CreateEndpoint', createEndpoint_username - The user name to be used to log in to the endpoint database.

$sel:extraConnectionAttributes:CreateEndpoint', createEndpoint_extraConnectionAttributes - Additional attributes associated with the connection. Each attribute is specified as a name-value pair associated by an equal sign (=). Multiple attributes are separated by a semicolon (;) with no additional white space. For information on the attributes available for connecting your source or target endpoint, see Working with DMS Endpoints in the Database Migration Service User Guide.

$sel:kafkaSettings:CreateEndpoint', createEndpoint_kafkaSettings - Settings in JSON format for the target Apache Kafka endpoint. For more information about the available settings, see Using object mapping to migrate data to a Kafka topic in the Database Migration Service User Guide.

$sel:docDbSettings:CreateEndpoint', createEndpoint_docDbSettings - Undocumented member.

$sel:dynamoDbSettings:CreateEndpoint', createEndpoint_dynamoDbSettings - Settings in JSON format for the target Amazon DynamoDB endpoint. For information about other available settings, see Using Object Mapping to Migrate Data to DynamoDB in the Database Migration Service User Guide.

$sel:kinesisSettings:CreateEndpoint', createEndpoint_kinesisSettings - Settings in JSON format for the target endpoint for Amazon Kinesis Data Streams. For more information about the available settings, see Using object mapping to migrate data to a Kinesis data stream in the Database Migration Service User Guide.

$sel:sybaseSettings:CreateEndpoint', createEndpoint_sybaseSettings - Settings in JSON format for the source and target SAP ASE endpoint. For information about other available settings, see Extra connection attributes when using SAP ASE as a source for DMS and Extra connection attributes when using SAP ASE as a target for DMS in the Database Migration Service User Guide.

$sel:databaseName:CreateEndpoint', createEndpoint_databaseName - The name of the endpoint database. For a MySQL source or target endpoint, do not specify DatabaseName.

$sel:endpointIdentifier:CreateEndpoint', createEndpoint_endpointIdentifier - The database endpoint identifier. Identifiers must begin with a letter and must contain only ASCII letters, digits, and hyphens. They can't end with a hyphen, or contain two consecutive hyphens.

$sel:endpointType:CreateEndpoint', createEndpoint_endpointType - The type of endpoint. Valid values are source and target.

$sel:engineName:CreateEndpoint', createEndpoint_engineName - The type of engine for the endpoint. Valid values, depending on the EndpointType value, include "mysql", "oracle", "postgres", "mariadb", "aurora", "aurora-postgresql", "redshift", "s3", "db2", "azuredb", "sybase", "dynamodb", "mongodb", "kinesis", "kafka", "elasticsearch", "docdb", "sqlserver", and "neptune".

Request Lenses

createEndpoint_mongoDbSettings :: Lens' CreateEndpoint (Maybe MongoDbSettings)

Settings in JSON format for the source MongoDB endpoint. For more information about the available settings, see Endpoint configuration settings when using MongoDB as a source for Database Migration Service in the Database Migration Service User Guide.

createEndpoint_sslMode :: Lens' CreateEndpoint (Maybe DmsSslModeValue)

The Secure Sockets Layer (SSL) mode to use for the SSL connection. The default is none

createEndpoint_neptuneSettings :: Lens' CreateEndpoint (Maybe NeptuneSettings)

Settings in JSON format for the target Amazon Neptune endpoint. For more information about the available settings, see Specifying graph-mapping rules using Gremlin and R2RML for Amazon Neptune as a target in the Database Migration Service User Guide.

createEndpoint_redisSettings :: Lens' CreateEndpoint (Maybe RedisSettings)

Settings in JSON format for the target Redis endpoint.

createEndpoint_externalTableDefinition :: Lens' CreateEndpoint (Maybe Text)

The external table definition.

createEndpoint_elasticsearchSettings :: Lens' CreateEndpoint (Maybe ElasticsearchSettings)

Settings in JSON format for the target Elasticsearch endpoint. For more information about the available settings, see Extra Connection Attributes When Using Elasticsearch as a Target for DMS in the Database Migration Service User Guide.

createEndpoint_postgreSQLSettings :: Lens' CreateEndpoint (Maybe PostgreSQLSettings)

Settings in JSON format for the source and target PostgreSQL endpoint. For information about other available settings, see Extra connection attributes when using PostgreSQL as a source for DMS and Extra connection attributes when using PostgreSQL as a target for DMS in the Database Migration Service User Guide.

createEndpoint_oracleSettings :: Lens' CreateEndpoint (Maybe OracleSettings)

Settings in JSON format for the source and target Oracle endpoint. For information about other available settings, see Extra connection attributes when using Oracle as a source for DMS and Extra connection attributes when using Oracle as a target for DMS in the Database Migration Service User Guide.

createEndpoint_certificateArn :: Lens' CreateEndpoint (Maybe Text)

The Amazon Resource Name (ARN) for the certificate.

createEndpoint_serviceAccessRoleArn :: Lens' CreateEndpoint (Maybe Text)

The Amazon Resource Name (ARN) for the service access role that you want to use to create the endpoint. The role must allow the iam:PassRole action.

createEndpoint_s3Settings :: Lens' CreateEndpoint (Maybe S3Settings)

Settings in JSON format for the target Amazon S3 endpoint. For more information about the available settings, see Extra Connection Attributes When Using Amazon S3 as a Target for DMS in the Database Migration Service User Guide.

createEndpoint_microsoftSQLServerSettings :: Lens' CreateEndpoint (Maybe MicrosoftSQLServerSettings)

Settings in JSON format for the source and target Microsoft SQL Server endpoint. For information about other available settings, see Extra connection attributes when using SQL Server as a source for DMS and Extra connection attributes when using SQL Server as a target for DMS in the Database Migration Service User Guide.

createEndpoint_serverName :: Lens' CreateEndpoint (Maybe Text)

The name of the server where the endpoint database resides.

createEndpoint_kmsKeyId :: Lens' CreateEndpoint (Maybe Text)

An KMS key identifier that is used to encrypt the connection parameters for the endpoint.

If you don't specify a value for the KmsKeyId parameter, then DMS uses your default encryption key.

KMS creates the default encryption key for your Amazon Web Services account. Your Amazon Web Services account has a different default encryption key for each Amazon Web Services Region.

createEndpoint_iBMDb2Settings :: Lens' CreateEndpoint (Maybe IBMDb2Settings)

Settings in JSON format for the source IBM Db2 LUW endpoint. For information about other available settings, see Extra connection attributes when using Db2 LUW as a source for DMS in the Database Migration Service User Guide.

createEndpoint_password :: Lens' CreateEndpoint (Maybe Text)

The password to be used to log in to the endpoint database.

createEndpoint_mySQLSettings :: Lens' CreateEndpoint (Maybe MySQLSettings)

Settings in JSON format for the source and target MySQL endpoint. For information about other available settings, see Extra connection attributes when using MySQL as a source for DMS and Extra connection attributes when using a MySQL-compatible database as a target for DMS in the Database Migration Service User Guide.

createEndpoint_dmsTransferSettings :: Lens' CreateEndpoint (Maybe DmsTransferSettings)

The settings in JSON format for the DMS transfer type of source endpoint.

Possible settings include the following:

ServiceAccessRoleArn - The Amazon Resource Name (ARN) used by the service access IAM role. The role must allow the iam:PassRole action.

BucketName - The name of the S3 bucket to use.

Shorthand syntax for these settings is as follows: ServiceAccessRoleArn=string,BucketName=string

JSON syntax for these settings is as follows: { "ServiceAccessRoleArn": "string", "BucketName": "string", }

createEndpoint_tags :: Lens' CreateEndpoint (Maybe [Tag])

One or more tags to be assigned to the endpoint.

createEndpoint_resourceIdentifier :: Lens' CreateEndpoint (Maybe Text)

A friendly name for the resource identifier at the end of the EndpointArn response parameter that is returned in the created Endpoint object. The value for this parameter can have up to 31 characters. It can contain only ASCII letters, digits, and hyphen ('-'). Also, it can't end with a hyphen or contain two consecutive hyphens, and can only begin with a letter, such as Example-App-ARN1. For example, this value might result in the EndpointArn value arn:aws:dms:eu-west-1:012345678901:rep:Example-App-ARN1. If you don't specify a ResourceIdentifier value, DMS generates a default identifier value for the end of EndpointArn.

createEndpoint_port :: Lens' CreateEndpoint (Maybe Int)

The port used by the endpoint database.

createEndpoint_redshiftSettings :: Lens' CreateEndpoint (Maybe RedshiftSettings)

Undocumented member.

createEndpoint_username :: Lens' CreateEndpoint (Maybe Text)

The user name to be used to log in to the endpoint database.

createEndpoint_extraConnectionAttributes :: Lens' CreateEndpoint (Maybe Text)

Additional attributes associated with the connection. Each attribute is specified as a name-value pair associated by an equal sign (=). Multiple attributes are separated by a semicolon (;) with no additional white space. For information on the attributes available for connecting your source or target endpoint, see Working with DMS Endpoints in the Database Migration Service User Guide.

createEndpoint_kafkaSettings :: Lens' CreateEndpoint (Maybe KafkaSettings)

Settings in JSON format for the target Apache Kafka endpoint. For more information about the available settings, see Using object mapping to migrate data to a Kafka topic in the Database Migration Service User Guide.

createEndpoint_docDbSettings :: Lens' CreateEndpoint (Maybe DocDbSettings)

Undocumented member.

createEndpoint_dynamoDbSettings :: Lens' CreateEndpoint (Maybe DynamoDbSettings)

Settings in JSON format for the target Amazon DynamoDB endpoint. For information about other available settings, see Using Object Mapping to Migrate Data to DynamoDB in the Database Migration Service User Guide.

createEndpoint_kinesisSettings :: Lens' CreateEndpoint (Maybe KinesisSettings)

Settings in JSON format for the target endpoint for Amazon Kinesis Data Streams. For more information about the available settings, see Using object mapping to migrate data to a Kinesis data stream in the Database Migration Service User Guide.

createEndpoint_sybaseSettings :: Lens' CreateEndpoint (Maybe SybaseSettings)

Settings in JSON format for the source and target SAP ASE endpoint. For information about other available settings, see Extra connection attributes when using SAP ASE as a source for DMS and Extra connection attributes when using SAP ASE as a target for DMS in the Database Migration Service User Guide.

createEndpoint_databaseName :: Lens' CreateEndpoint (Maybe Text)

The name of the endpoint database. For a MySQL source or target endpoint, do not specify DatabaseName.

createEndpoint_endpointIdentifier :: Lens' CreateEndpoint Text

The database endpoint identifier. Identifiers must begin with a letter and must contain only ASCII letters, digits, and hyphens. They can't end with a hyphen, or contain two consecutive hyphens.

createEndpoint_endpointType :: Lens' CreateEndpoint ReplicationEndpointTypeValue

The type of endpoint. Valid values are source and target.

createEndpoint_engineName :: Lens' CreateEndpoint Text

The type of engine for the endpoint. Valid values, depending on the EndpointType value, include "mysql", "oracle", "postgres", "mariadb", "aurora", "aurora-postgresql", "redshift", "s3", "db2", "azuredb", "sybase", "dynamodb", "mongodb", "kinesis", "kafka", "elasticsearch", "docdb", "sqlserver", and "neptune".

Destructuring the Response

data CreateEndpointResponse

See: newCreateEndpointResponse smart constructor.

Constructors

CreateEndpointResponse'

Fields

endpoint :: Maybe Endpoint

The endpoint that was created.

httpStatus :: Int

The response's http status code.

newCreateEndpointResponse

Arguments

:: Int
$sel:httpStatus:CreateEndpointResponse'

-> CreateEndpointResponse

Create a value of CreateEndpointResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:endpoint:CreateEndpointResponse', createEndpointResponse_endpoint - The endpoint that was created.

$sel:httpStatus:CreateEndpointResponse', createEndpointResponse_httpStatus - The response's http status code.

Response Lenses

createEndpointResponse_endpoint :: Lens' CreateEndpointResponse (Maybe Endpoint)

The endpoint that was created.

createEndpointResponse_httpStatus :: Lens' CreateEndpointResponse Int

The response's http status code.