Network.AWS.DMS.CreateEventSubscription

Description

Creates an DMS event notification subscription.

You can specify the type of source (SourceType) you want to be notified of, provide a list of DMS source IDs (SourceIds) that triggers the events, and provide a list of event categories (EventCategories) for events you want to be notified of. If you specify both the SourceType and SourceIds, such as SourceType = replication-instance and SourceIdentifier = my-replinstance, you will be notified of all the replication instance events for the specified source. If you specify a SourceType but don't specify a SourceIdentifier, you receive notice of the events for that source type for all your DMS sources. If you don't specify either SourceType nor SourceIdentifier, you will be notified of events generated from all DMS sources belonging to your customer account.

For more information about DMS events, see Working with Events and Notifications in the Database Migration Service User Guide.

Creating a Request

data CreateEventSubscription

See: newCreateEventSubscription smart constructor.

Constructors

CreateEventSubscription'

Fields

enabled :: Maybe Bool

A Boolean value; set to true to activate the subscription, or set to false to create the subscription but not activate it.

sourceType :: Maybe Text

The type of DMS resource that generates the events. For example, if you want to be notified of events generated by a replication instance, you set this parameter to replication-instance. If this value isn't specified, all events are returned.

Valid values: replication-instance | replication-task

eventCategories :: Maybe [Text]

A list of event categories for a source type that you want to subscribe to. For more information, see Working with Events and Notifications in the Database Migration Service User Guide.

sourceIds :: Maybe [Text]

A list of identifiers for which DMS provides notification events.

If you don't specify a value, notifications are provided for all sources.

If you specify multiple values, they must be of the same type. For example, if you specify a database instance ID, then all of the other values must be database instance IDs.

tags :: Maybe [Tag]

One or more tags to be assigned to the event subscription.

subscriptionName :: Text

The name of the DMS event notification subscription. This name must be less than 255 characters.

snsTopicArn :: Text

The Amazon Resource Name (ARN) of the Amazon SNS topic created for event notification. The ARN is created by Amazon SNS when you create a topic and subscribe to it.

newCreateEventSubscription

Arguments

:: Text
$sel:subscriptionName:CreateEventSubscription'

-> Text
$sel:snsTopicArn:CreateEventSubscription'

-> CreateEventSubscription

Create a value of CreateEventSubscription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:enabled:CreateEventSubscription', createEventSubscription_enabled - A Boolean value; set to true to activate the subscription, or set to false to create the subscription but not activate it.

$sel:sourceType:CreateEventSubscription', createEventSubscription_sourceType - The type of DMS resource that generates the events. For example, if you want to be notified of events generated by a replication instance, you set this parameter to replication-instance. If this value isn't specified, all events are returned.

Valid values: replication-instance | replication-task

$sel:eventCategories:CreateEventSubscription', createEventSubscription_eventCategories - A list of event categories for a source type that you want to subscribe to. For more information, see Working with Events and Notifications in the Database Migration Service User Guide.

$sel:sourceIds:CreateEventSubscription', createEventSubscription_sourceIds - A list of identifiers for which DMS provides notification events.

If you don't specify a value, notifications are provided for all sources.

If you specify multiple values, they must be of the same type. For example, if you specify a database instance ID, then all of the other values must be database instance IDs.

$sel:tags:CreateEventSubscription', createEventSubscription_tags - One or more tags to be assigned to the event subscription.

$sel:subscriptionName:CreateEventSubscription', createEventSubscription_subscriptionName - The name of the DMS event notification subscription. This name must be less than 255 characters.

$sel:snsTopicArn:CreateEventSubscription', createEventSubscription_snsTopicArn - The Amazon Resource Name (ARN) of the Amazon SNS topic created for event notification. The ARN is created by Amazon SNS when you create a topic and subscribe to it.

Request Lenses

createEventSubscription_enabled :: Lens' CreateEventSubscription (Maybe Bool)

A Boolean value; set to true to activate the subscription, or set to false to create the subscription but not activate it.

createEventSubscription_sourceType :: Lens' CreateEventSubscription (Maybe Text)

The type of DMS resource that generates the events. For example, if you want to be notified of events generated by a replication instance, you set this parameter to replication-instance. If this value isn't specified, all events are returned.

Valid values: replication-instance | replication-task

createEventSubscription_eventCategories :: Lens' CreateEventSubscription (Maybe [Text])

A list of event categories for a source type that you want to subscribe to. For more information, see Working with Events and Notifications in the Database Migration Service User Guide.

createEventSubscription_sourceIds :: Lens' CreateEventSubscription (Maybe [Text])

A list of identifiers for which DMS provides notification events.

If you don't specify a value, notifications are provided for all sources.

If you specify multiple values, they must be of the same type. For example, if you specify a database instance ID, then all of the other values must be database instance IDs.

createEventSubscription_tags :: Lens' CreateEventSubscription (Maybe [Tag])

One or more tags to be assigned to the event subscription.

createEventSubscription_subscriptionName :: Lens' CreateEventSubscription Text

The name of the DMS event notification subscription. This name must be less than 255 characters.

createEventSubscription_snsTopicArn :: Lens' CreateEventSubscription Text

The Amazon Resource Name (ARN) of the Amazon SNS topic created for event notification. The ARN is created by Amazon SNS when you create a topic and subscribe to it.

Destructuring the Response

data CreateEventSubscriptionResponse

See: newCreateEventSubscriptionResponse smart constructor.

Constructors

CreateEventSubscriptionResponse'

Fields

eventSubscription :: Maybe EventSubscription

The event subscription that was created.

httpStatus :: Int

The response's http status code.

newCreateEventSubscriptionResponse

Arguments

:: Int
$sel:httpStatus:CreateEventSubscriptionResponse'

-> CreateEventSubscriptionResponse

Create a value of CreateEventSubscriptionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventSubscription:CreateEventSubscriptionResponse', createEventSubscriptionResponse_eventSubscription - The event subscription that was created.

$sel:httpStatus:CreateEventSubscriptionResponse', createEventSubscriptionResponse_httpStatus - The response's http status code.

Response Lenses

createEventSubscriptionResponse_eventSubscription :: Lens' CreateEventSubscriptionResponse (Maybe EventSubscription)

The event subscription that was created.

createEventSubscriptionResponse_httpStatus :: Lens' CreateEventSubscriptionResponse Int

The response's http status code.