Network.AWS.DMS.CreateReplicationInstance

Description

Creates the replication instance using the specified parameters.

DMS requires that your account have certain roles with appropriate permissions before you can create a replication instance. For information on the required roles, see Creating the IAM Roles to Use With the CLI and DMS API. For information on the required permissions, see IAM Permissions Needed to Use DMS.

Creating a Request

data CreateReplicationInstance

See: newCreateReplicationInstance smart constructor.

Constructors

CreateReplicationInstance'

Fields

replicationSubnetGroupIdentifier :: Maybe Text

A subnet group to associate with the replication instance.

publiclyAccessible :: Maybe Bool

Specifies the accessibility options for the replication instance. A value of true represents an instance with a public IP address. A value of false represents an instance with a private IP address. The default value is true.

multiAZ :: Maybe Bool

Specifies whether the replication instance is a Multi-AZ deployment. You can't set the AvailabilityZone parameter if the Multi-AZ parameter is set to true.

vpcSecurityGroupIds :: Maybe [Text]

Specifies the VPC security group to be used with the replication instance. The VPC security group must work with the VPC containing the replication instance.

kmsKeyId :: Maybe Text

An KMS key identifier that is used to encrypt the data on the replication instance.

If you don't specify a value for the KmsKeyId parameter, then DMS uses your default encryption key.

KMS creates the default encryption key for your Amazon Web Services account. Your Amazon Web Services account has a different default encryption key for each Amazon Web Services Region.

availabilityZone :: Maybe Text

The Availability Zone where the replication instance will be created. The default value is a random, system-chosen Availability Zone in the endpoint's Amazon Web Services Region, for example: us-east-1d

engineVersion :: Maybe Text

The engine version number of the replication instance.

If an engine version number is not specified when a replication instance is created, the default is the latest engine version available.

preferredMaintenanceWindow :: Maybe Text

The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

Format: ddd:hh24:mi-ddd:hh24:mi

Default: A 30-minute window selected at random from an 8-hour block of time per Amazon Web Services Region, occurring on a random day of the week.

Valid Days: Mon, Tue, Wed, Thu, Fri, Sat, Sun

Constraints: Minimum 30-minute window.

tags :: Maybe [Tag]

One or more tags to be assigned to the replication instance.

resourceIdentifier :: Maybe Text

A friendly name for the resource identifier at the end of the EndpointArn response parameter that is returned in the created Endpoint object. The value for this parameter can have up to 31 characters. It can contain only ASCII letters, digits, and hyphen ('-'). Also, it can't end with a hyphen or contain two consecutive hyphens, and can only begin with a letter, such as Example-App-ARN1. For example, this value might result in the EndpointArn value arn:aws:dms:eu-west-1:012345678901:rep:Example-App-ARN1. If you don't specify a ResourceIdentifier value, DMS generates a default identifier value for the end of EndpointArn.

dnsNameServers :: Maybe Text

A list of custom DNS name servers supported for the replication instance to access your on-premise source or target database. This list overrides the default name servers supported by the replication instance. You can specify a comma-separated list of internet addresses for up to four on-premise DNS name servers. For example: "1.1.1.1,2.2.2.2,3.3.3.3,4.4.4.4"

allocatedStorage :: Maybe Int

The amount of storage (in gigabytes) to be initially allocated for the replication instance.

autoMinorVersionUpgrade :: Maybe Bool

A value that indicates whether minor engine upgrades are applied automatically to the replication instance during the maintenance window. This parameter defaults to true.

Default: true

replicationInstanceIdentifier :: Text

The replication instance identifier. This parameter is stored as a lowercase string.

Constraints:

Must contain 1-63 alphanumeric characters or hyphens.

First character must be a letter.

Can't end with a hyphen or contain two consecutive hyphens.

Example: myrepinstance

replicationInstanceClass :: Text

The compute and memory capacity of the replication instance as defined for the specified replication instance class. For example to specify the instance class dms.c4.large, set this parameter to "dms.c4.large".

For more information on the settings and capacities for the available replication instance classes, see Selecting the right DMS replication instance for your migration.

newCreateReplicationInstance

Arguments

:: Text
$sel:replicationInstanceIdentifier:CreateReplicationInstance'

-> Text
$sel:replicationInstanceClass:CreateReplicationInstance'

-> CreateReplicationInstance

Create a value of CreateReplicationInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:replicationSubnetGroupIdentifier:CreateReplicationInstance', createReplicationInstance_replicationSubnetGroupIdentifier - A subnet group to associate with the replication instance.

$sel:publiclyAccessible:CreateReplicationInstance', createReplicationInstance_publiclyAccessible - Specifies the accessibility options for the replication instance. A value of true represents an instance with a public IP address. A value of false represents an instance with a private IP address. The default value is true.

$sel:multiAZ:CreateReplicationInstance', createReplicationInstance_multiAZ - Specifies whether the replication instance is a Multi-AZ deployment. You can't set the AvailabilityZone parameter if the Multi-AZ parameter is set to true.

$sel:vpcSecurityGroupIds:CreateReplicationInstance', createReplicationInstance_vpcSecurityGroupIds - Specifies the VPC security group to be used with the replication instance. The VPC security group must work with the VPC containing the replication instance.

$sel:kmsKeyId:CreateReplicationInstance', createReplicationInstance_kmsKeyId - An KMS key identifier that is used to encrypt the data on the replication instance.

If you don't specify a value for the KmsKeyId parameter, then DMS uses your default encryption key.

KMS creates the default encryption key for your Amazon Web Services account. Your Amazon Web Services account has a different default encryption key for each Amazon Web Services Region.

$sel:availabilityZone:CreateReplicationInstance', createReplicationInstance_availabilityZone - The Availability Zone where the replication instance will be created. The default value is a random, system-chosen Availability Zone in the endpoint's Amazon Web Services Region, for example: us-east-1d

$sel:engineVersion:CreateReplicationInstance', createReplicationInstance_engineVersion - The engine version number of the replication instance.

If an engine version number is not specified when a replication instance is created, the default is the latest engine version available.

$sel:preferredMaintenanceWindow:CreateReplicationInstance', createReplicationInstance_preferredMaintenanceWindow - The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

Format: ddd:hh24:mi-ddd:hh24:mi

Default: A 30-minute window selected at random from an 8-hour block of time per Amazon Web Services Region, occurring on a random day of the week.

Valid Days: Mon, Tue, Wed, Thu, Fri, Sat, Sun

Constraints: Minimum 30-minute window.

$sel:tags:CreateReplicationInstance', createReplicationInstance_tags - One or more tags to be assigned to the replication instance.

$sel:resourceIdentifier:CreateReplicationInstance', createReplicationInstance_resourceIdentifier - A friendly name for the resource identifier at the end of the EndpointArn response parameter that is returned in the created Endpoint object. The value for this parameter can have up to 31 characters. It can contain only ASCII letters, digits, and hyphen ('-'). Also, it can't end with a hyphen or contain two consecutive hyphens, and can only begin with a letter, such as Example-App-ARN1. For example, this value might result in the EndpointArn value arn:aws:dms:eu-west-1:012345678901:rep:Example-App-ARN1. If you don't specify a ResourceIdentifier value, DMS generates a default identifier value for the end of EndpointArn.

$sel:dnsNameServers:CreateReplicationInstance', createReplicationInstance_dnsNameServers - A list of custom DNS name servers supported for the replication instance to access your on-premise source or target database. This list overrides the default name servers supported by the replication instance. You can specify a comma-separated list of internet addresses for up to four on-premise DNS name servers. For example: "1.1.1.1,2.2.2.2,3.3.3.3,4.4.4.4"

$sel:allocatedStorage:CreateReplicationInstance', createReplicationInstance_allocatedStorage - The amount of storage (in gigabytes) to be initially allocated for the replication instance.

$sel:autoMinorVersionUpgrade:CreateReplicationInstance', createReplicationInstance_autoMinorVersionUpgrade - A value that indicates whether minor engine upgrades are applied automatically to the replication instance during the maintenance window. This parameter defaults to true.

Default: true

$sel:replicationInstanceIdentifier:CreateReplicationInstance', createReplicationInstance_replicationInstanceIdentifier - The replication instance identifier. This parameter is stored as a lowercase string.

Constraints:

Must contain 1-63 alphanumeric characters or hyphens.

First character must be a letter.

Can't end with a hyphen or contain two consecutive hyphens.

Example: myrepinstance

$sel:replicationInstanceClass:CreateReplicationInstance', createReplicationInstance_replicationInstanceClass - The compute and memory capacity of the replication instance as defined for the specified replication instance class. For example to specify the instance class dms.c4.large, set this parameter to "dms.c4.large".

For more information on the settings and capacities for the available replication instance classes, see Selecting the right DMS replication instance for your migration.

Request Lenses

createReplicationInstance_replicationSubnetGroupIdentifier :: Lens' CreateReplicationInstance (Maybe Text)

A subnet group to associate with the replication instance.

createReplicationInstance_publiclyAccessible :: Lens' CreateReplicationInstance (Maybe Bool)

Specifies the accessibility options for the replication instance. A value of true represents an instance with a public IP address. A value of false represents an instance with a private IP address. The default value is true.

createReplicationInstance_multiAZ :: Lens' CreateReplicationInstance (Maybe Bool)

Specifies whether the replication instance is a Multi-AZ deployment. You can't set the AvailabilityZone parameter if the Multi-AZ parameter is set to true.

createReplicationInstance_vpcSecurityGroupIds :: Lens' CreateReplicationInstance (Maybe [Text])

Specifies the VPC security group to be used with the replication instance. The VPC security group must work with the VPC containing the replication instance.

createReplicationInstance_kmsKeyId :: Lens' CreateReplicationInstance (Maybe Text)

An KMS key identifier that is used to encrypt the data on the replication instance.

If you don't specify a value for the KmsKeyId parameter, then DMS uses your default encryption key.

KMS creates the default encryption key for your Amazon Web Services account. Your Amazon Web Services account has a different default encryption key for each Amazon Web Services Region.

createReplicationInstance_availabilityZone :: Lens' CreateReplicationInstance (Maybe Text)

The Availability Zone where the replication instance will be created. The default value is a random, system-chosen Availability Zone in the endpoint's Amazon Web Services Region, for example: us-east-1d

createReplicationInstance_engineVersion :: Lens' CreateReplicationInstance (Maybe Text)

The engine version number of the replication instance.

If an engine version number is not specified when a replication instance is created, the default is the latest engine version available.

createReplicationInstance_preferredMaintenanceWindow :: Lens' CreateReplicationInstance (Maybe Text)

The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

Format: ddd:hh24:mi-ddd:hh24:mi

Default: A 30-minute window selected at random from an 8-hour block of time per Amazon Web Services Region, occurring on a random day of the week.

Valid Days: Mon, Tue, Wed, Thu, Fri, Sat, Sun

Constraints: Minimum 30-minute window.

createReplicationInstance_tags :: Lens' CreateReplicationInstance (Maybe [Tag])

One or more tags to be assigned to the replication instance.

createReplicationInstance_resourceIdentifier :: Lens' CreateReplicationInstance (Maybe Text)

A friendly name for the resource identifier at the end of the EndpointArn response parameter that is returned in the created Endpoint object. The value for this parameter can have up to 31 characters. It can contain only ASCII letters, digits, and hyphen ('-'). Also, it can't end with a hyphen or contain two consecutive hyphens, and can only begin with a letter, such as Example-App-ARN1. For example, this value might result in the EndpointArn value arn:aws:dms:eu-west-1:012345678901:rep:Example-App-ARN1. If you don't specify a ResourceIdentifier value, DMS generates a default identifier value for the end of EndpointArn.

createReplicationInstance_dnsNameServers :: Lens' CreateReplicationInstance (Maybe Text)

A list of custom DNS name servers supported for the replication instance to access your on-premise source or target database. This list overrides the default name servers supported by the replication instance. You can specify a comma-separated list of internet addresses for up to four on-premise DNS name servers. For example: "1.1.1.1,2.2.2.2,3.3.3.3,4.4.4.4"

createReplicationInstance_allocatedStorage :: Lens' CreateReplicationInstance (Maybe Int)

The amount of storage (in gigabytes) to be initially allocated for the replication instance.

createReplicationInstance_autoMinorVersionUpgrade :: Lens' CreateReplicationInstance (Maybe Bool)

A value that indicates whether minor engine upgrades are applied automatically to the replication instance during the maintenance window. This parameter defaults to true.

Default: true

createReplicationInstance_replicationInstanceIdentifier :: Lens' CreateReplicationInstance Text

The replication instance identifier. This parameter is stored as a lowercase string.

Constraints:

Must contain 1-63 alphanumeric characters or hyphens.

First character must be a letter.

Can't end with a hyphen or contain two consecutive hyphens.

Example: myrepinstance

createReplicationInstance_replicationInstanceClass :: Lens' CreateReplicationInstance Text

The compute and memory capacity of the replication instance as defined for the specified replication instance class. For example to specify the instance class dms.c4.large, set this parameter to "dms.c4.large".

For more information on the settings and capacities for the available replication instance classes, see Selecting the right DMS replication instance for your migration.

Destructuring the Response

data CreateReplicationInstanceResponse

See: newCreateReplicationInstanceResponse smart constructor.

Constructors

CreateReplicationInstanceResponse'

Fields

replicationInstance :: Maybe ReplicationInstance

The replication instance that was created.

httpStatus :: Int

The response's http status code.

newCreateReplicationInstanceResponse

Arguments

:: Int
$sel:httpStatus:CreateReplicationInstanceResponse'

-> CreateReplicationInstanceResponse

Create a value of CreateReplicationInstanceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:replicationInstance:CreateReplicationInstanceResponse', createReplicationInstanceResponse_replicationInstance - The replication instance that was created.

$sel:httpStatus:CreateReplicationInstanceResponse', createReplicationInstanceResponse_httpStatus - The response's http status code.

Response Lenses

createReplicationInstanceResponse_replicationInstance :: Lens' CreateReplicationInstanceResponse (Maybe ReplicationInstance)

The replication instance that was created.

createReplicationInstanceResponse_httpStatus :: Lens' CreateReplicationInstanceResponse Int

The response's http status code.