Network.AWS.DMS.CreateReplicationSubnetGroup

Description

Creates a replication subnet group given a list of the subnet IDs in a VPC.

The VPC needs to have at least one subnet in at least two availability zones in the Amazon Web Services Region, otherwise the service will throw a ReplicationSubnetGroupDoesNotCoverEnoughAZs exception.

Creating a Request

data CreateReplicationSubnetGroup

See: newCreateReplicationSubnetGroup smart constructor.

Constructors

CreateReplicationSubnetGroup'

Fields

tags :: Maybe [Tag]

One or more tags to be assigned to the subnet group.

replicationSubnetGroupIdentifier :: Text

The name for the replication subnet group. This value is stored as a lowercase string.

Constraints: Must contain no more than 255 alphanumeric characters, periods, spaces, underscores, or hyphens. Must not be "default".

Example: mySubnetgroup

replicationSubnetGroupDescription :: Text

The description for the subnet group.

subnetIds :: [Text]

One or more subnet IDs to be assigned to the subnet group.

newCreateReplicationSubnetGroup

Arguments

:: Text
$sel:replicationSubnetGroupIdentifier:CreateReplicationSubnetGroup'

-> Text
$sel:replicationSubnetGroupDescription:CreateReplicationSubnetGroup'

-> CreateReplicationSubnetGroup

Create a value of CreateReplicationSubnetGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateReplicationSubnetGroup', createReplicationSubnetGroup_tags - One or more tags to be assigned to the subnet group.

$sel:replicationSubnetGroupIdentifier:CreateReplicationSubnetGroup', createReplicationSubnetGroup_replicationSubnetGroupIdentifier - The name for the replication subnet group. This value is stored as a lowercase string.

Constraints: Must contain no more than 255 alphanumeric characters, periods, spaces, underscores, or hyphens. Must not be "default".

Example: mySubnetgroup

$sel:replicationSubnetGroupDescription:CreateReplicationSubnetGroup', createReplicationSubnetGroup_replicationSubnetGroupDescription - The description for the subnet group.

$sel:subnetIds:CreateReplicationSubnetGroup', createReplicationSubnetGroup_subnetIds - One or more subnet IDs to be assigned to the subnet group.

Request Lenses

createReplicationSubnetGroup_tags :: Lens' CreateReplicationSubnetGroup (Maybe [Tag])

One or more tags to be assigned to the subnet group.

createReplicationSubnetGroup_replicationSubnetGroupIdentifier :: Lens' CreateReplicationSubnetGroup Text

The name for the replication subnet group. This value is stored as a lowercase string.

Constraints: Must contain no more than 255 alphanumeric characters, periods, spaces, underscores, or hyphens. Must not be "default".

Example: mySubnetgroup

createReplicationSubnetGroup_replicationSubnetGroupDescription :: Lens' CreateReplicationSubnetGroup Text

The description for the subnet group.

createReplicationSubnetGroup_subnetIds :: Lens' CreateReplicationSubnetGroup [Text]

One or more subnet IDs to be assigned to the subnet group.

Destructuring the Response

data CreateReplicationSubnetGroupResponse

See: newCreateReplicationSubnetGroupResponse smart constructor.

Constructors

CreateReplicationSubnetGroupResponse'

Fields

replicationSubnetGroup :: Maybe ReplicationSubnetGroup

The replication subnet group that was created.

httpStatus :: Int

The response's http status code.

newCreateReplicationSubnetGroupResponse

Arguments

:: Int
$sel:httpStatus:CreateReplicationSubnetGroupResponse'

-> CreateReplicationSubnetGroupResponse

Create a value of CreateReplicationSubnetGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:replicationSubnetGroup:CreateReplicationSubnetGroupResponse', createReplicationSubnetGroupResponse_replicationSubnetGroup - The replication subnet group that was created.

$sel:httpStatus:CreateReplicationSubnetGroupResponse', createReplicationSubnetGroupResponse_httpStatus - The response's http status code.

Response Lenses

createReplicationSubnetGroupResponse_replicationSubnetGroup :: Lens' CreateReplicationSubnetGroupResponse (Maybe ReplicationSubnetGroup)

The replication subnet group that was created.

createReplicationSubnetGroupResponse_httpStatus :: Lens' CreateReplicationSubnetGroupResponse Int

The response's http status code.