Network.AWS.DMS.DescribeApplicableIndividualAssessments

Description

Provides a list of individual assessments that you can specify for a new premigration assessment run, given one or more parameters.

If you specify an existing migration task, this operation provides the default individual assessments you can specify for that task. Otherwise, the specified parameters model elements of a possible migration task on which to base a premigration assessment run.

To use these migration task modeling parameters, you must specify an existing replication instance, a source database engine, a target database engine, and a migration type. This combination of parameters potentially limits the default individual assessments available for an assessment run created for a corresponding migration task.

If you specify no parameters, this operation provides a list of all possible individual assessments that you can specify for an assessment run. If you specify any one of the task modeling parameters, you must specify all of them or the operation cannot provide a list of individual assessments. The only parameter that you can specify alone is for an existing migration task. The specified task definition then determines the default list of individual assessments that you can specify in an assessment run for the task.

Creating a Request

data DescribeApplicableIndividualAssessments

See: newDescribeApplicableIndividualAssessments smart constructor.

Constructors

DescribeApplicableIndividualAssessments'

Fields

migrationType :: Maybe MigrationTypeValue

Name of the migration type that each provided individual assessment must support.

sourceEngineName :: Maybe Text

Name of a database engine that the specified replication instance supports as a source.

replicationTaskArn :: Maybe Text

Amazon Resource Name (ARN) of a migration task on which you want to base the default list of individual assessments.

marker :: Maybe Text

Optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

maxRecords :: Maybe Int

Maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

targetEngineName :: Maybe Text

Name of a database engine that the specified replication instance supports as a target.

replicationInstanceArn :: Maybe Text

ARN of a replication instance on which you want to base the default list of individual assessments.

newDescribeApplicableIndividualAssessments :: DescribeApplicableIndividualAssessments

Create a value of DescribeApplicableIndividualAssessments with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:migrationType:DescribeApplicableIndividualAssessments', describeApplicableIndividualAssessments_migrationType - Name of the migration type that each provided individual assessment must support.

$sel:sourceEngineName:DescribeApplicableIndividualAssessments', describeApplicableIndividualAssessments_sourceEngineName - Name of a database engine that the specified replication instance supports as a source.

$sel:replicationTaskArn:DescribeApplicableIndividualAssessments', describeApplicableIndividualAssessments_replicationTaskArn - Amazon Resource Name (ARN) of a migration task on which you want to base the default list of individual assessments.

$sel:marker:DescribeApplicableIndividualAssessments', describeApplicableIndividualAssessments_marker - Optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

$sel:maxRecords:DescribeApplicableIndividualAssessments', describeApplicableIndividualAssessments_maxRecords - Maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

$sel:targetEngineName:DescribeApplicableIndividualAssessments', describeApplicableIndividualAssessments_targetEngineName - Name of a database engine that the specified replication instance supports as a target.

$sel:replicationInstanceArn:DescribeApplicableIndividualAssessments', describeApplicableIndividualAssessments_replicationInstanceArn - ARN of a replication instance on which you want to base the default list of individual assessments.

Request Lenses

describeApplicableIndividualAssessments_migrationType :: Lens' DescribeApplicableIndividualAssessments (Maybe MigrationTypeValue)

Name of the migration type that each provided individual assessment must support.

describeApplicableIndividualAssessments_sourceEngineName :: Lens' DescribeApplicableIndividualAssessments (Maybe Text)

Name of a database engine that the specified replication instance supports as a source.

describeApplicableIndividualAssessments_replicationTaskArn :: Lens' DescribeApplicableIndividualAssessments (Maybe Text)

Amazon Resource Name (ARN) of a migration task on which you want to base the default list of individual assessments.

describeApplicableIndividualAssessments_marker :: Lens' DescribeApplicableIndividualAssessments (Maybe Text)

Optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

describeApplicableIndividualAssessments_maxRecords :: Lens' DescribeApplicableIndividualAssessments (Maybe Int)

Maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

describeApplicableIndividualAssessments_targetEngineName :: Lens' DescribeApplicableIndividualAssessments (Maybe Text)

Name of a database engine that the specified replication instance supports as a target.

describeApplicableIndividualAssessments_replicationInstanceArn :: Lens' DescribeApplicableIndividualAssessments (Maybe Text)

ARN of a replication instance on which you want to base the default list of individual assessments.

Destructuring the Response

data DescribeApplicableIndividualAssessmentsResponse

See: newDescribeApplicableIndividualAssessmentsResponse smart constructor.

Constructors

DescribeApplicableIndividualAssessmentsResponse'

Fields

marker :: Maybe Text

Pagination token returned for you to pass to a subsequent request. If you pass this token as the Marker value in a subsequent request, the response includes only records beyond the marker, up to the value specified in the request by MaxRecords.

individualAssessmentNames :: Maybe [Text]

List of names for the individual assessments supported by the premigration assessment run that you start based on the specified request parameters. For more information on the available individual assessments, including compatibility with different migration task configurations, see Working with premigration assessment runs in the Database Migration Service User Guide.

httpStatus :: Int

The response's http status code.

newDescribeApplicableIndividualAssessmentsResponse

Arguments

:: Int
$sel:httpStatus:DescribeApplicableIndividualAssessmentsResponse'

-> DescribeApplicableIndividualAssessmentsResponse

Create a value of DescribeApplicableIndividualAssessmentsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:marker:DescribeApplicableIndividualAssessments', describeApplicableIndividualAssessmentsResponse_marker - Pagination token returned for you to pass to a subsequent request. If you pass this token as the Marker value in a subsequent request, the response includes only records beyond the marker, up to the value specified in the request by MaxRecords.

$sel:individualAssessmentNames:DescribeApplicableIndividualAssessmentsResponse', describeApplicableIndividualAssessmentsResponse_individualAssessmentNames - List of names for the individual assessments supported by the premigration assessment run that you start based on the specified request parameters. For more information on the available individual assessments, including compatibility with different migration task configurations, see Working with premigration assessment runs in the Database Migration Service User Guide.

$sel:httpStatus:DescribeApplicableIndividualAssessmentsResponse', describeApplicableIndividualAssessmentsResponse_httpStatus - The response's http status code.

Response Lenses

describeApplicableIndividualAssessmentsResponse_marker :: Lens' DescribeApplicableIndividualAssessmentsResponse (Maybe Text)

Pagination token returned for you to pass to a subsequent request. If you pass this token as the Marker value in a subsequent request, the response includes only records beyond the marker, up to the value specified in the request by MaxRecords.

describeApplicableIndividualAssessmentsResponse_individualAssessmentNames :: Lens' DescribeApplicableIndividualAssessmentsResponse (Maybe [Text])

List of names for the individual assessments supported by the premigration assessment run that you start based on the specified request parameters. For more information on the available individual assessments, including compatibility with different migration task configurations, see Working with premigration assessment runs in the Database Migration Service User Guide.

describeApplicableIndividualAssessmentsResponse_httpStatus :: Lens' DescribeApplicableIndividualAssessmentsResponse Int

The response's http status code.