Network.AWS.DMS.DescribeConnections

Description

Describes the status of the connections that have been made between the replication instance and an endpoint. Connections are created when you test an endpoint.

This operation returns paginated results.

Creating a Request

data DescribeConnections

See: newDescribeConnections smart constructor.

Constructors

DescribeConnections'

Fields

filters :: Maybe [Filter]

The filters applied to the connection.

Valid filter names: endpoint-arn | replication-instance-arn

marker :: Maybe Text

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

maxRecords :: Maybe Int

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.

newDescribeConnections :: DescribeConnections

Create a value of DescribeConnections with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:DescribeConnections', describeConnections_filters - The filters applied to the connection.

Valid filter names: endpoint-arn | replication-instance-arn

$sel:marker:DescribeConnections', describeConnections_marker - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

$sel:maxRecords:DescribeConnections', describeConnections_maxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.

Request Lenses

describeConnections_filters :: Lens' DescribeConnections (Maybe [Filter])

The filters applied to the connection.

Valid filter names: endpoint-arn | replication-instance-arn

describeConnections_marker :: Lens' DescribeConnections (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

describeConnections_maxRecords :: Lens' DescribeConnections (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.

Destructuring the Response

data DescribeConnectionsResponse

See: newDescribeConnectionsResponse smart constructor.

Constructors

DescribeConnectionsResponse'

Fields

connections :: Maybe [Connection]

A description of the connections.

marker :: Maybe Text

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

httpStatus :: Int

The response's http status code.

newDescribeConnectionsResponse

Arguments

:: Int
$sel:httpStatus:DescribeConnectionsResponse'

-> DescribeConnectionsResponse

Create a value of DescribeConnectionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:connections:DescribeConnectionsResponse', describeConnectionsResponse_connections - A description of the connections.

$sel:marker:DescribeConnections', describeConnectionsResponse_marker - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

$sel:httpStatus:DescribeConnectionsResponse', describeConnectionsResponse_httpStatus - The response's http status code.

Response Lenses

describeConnectionsResponse_connections :: Lens' DescribeConnectionsResponse (Maybe [Connection])

A description of the connections.

describeConnectionsResponse_marker :: Lens' DescribeConnectionsResponse (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

describeConnectionsResponse_httpStatus :: Lens' DescribeConnectionsResponse Int

The response's http status code.