Network.AWS.DMS.DescribeEndpointTypes

Description

Returns information about the type of endpoints available.

This operation returns paginated results.

Creating a Request

data DescribeEndpointTypes

See: newDescribeEndpointTypes smart constructor.

Constructors

DescribeEndpointTypes'

Fields

filters :: Maybe [Filter]

Filters applied to the endpoint types.

Valid filter names: engine-name | endpoint-type

maxRecords :: Maybe Int

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.

marker :: Maybe Text

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

newDescribeEndpointTypes :: DescribeEndpointTypes

Create a value of DescribeEndpointTypes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:DescribeEndpointTypes', describeEndpointTypes_filters - Filters applied to the endpoint types.

Valid filter names: engine-name | endpoint-type

$sel:maxRecords:DescribeEndpointTypes', describeEndpointTypes_maxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.

$sel:marker:DescribeEndpointTypes', describeEndpointTypes_marker - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Request Lenses

describeEndpointTypes_filters :: Lens' DescribeEndpointTypes (Maybe [Filter])

Filters applied to the endpoint types.

Valid filter names: engine-name | endpoint-type

describeEndpointTypes_maxRecords :: Lens' DescribeEndpointTypes (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.

describeEndpointTypes_marker :: Lens' DescribeEndpointTypes (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Destructuring the Response

data DescribeEndpointTypesResponse

See: newDescribeEndpointTypesResponse smart constructor.

Constructors

DescribeEndpointTypesResponse'

Fields

supportedEndpointTypes :: Maybe [SupportedEndpointType]

The types of endpoints that are supported.

marker :: Maybe Text

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

httpStatus :: Int

The response's http status code.

newDescribeEndpointTypesResponse

Arguments

:: Int
$sel:httpStatus:DescribeEndpointTypesResponse'

-> DescribeEndpointTypesResponse

Create a value of DescribeEndpointTypesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:supportedEndpointTypes:DescribeEndpointTypesResponse', describeEndpointTypesResponse_supportedEndpointTypes - The types of endpoints that are supported.

$sel:marker:DescribeEndpointTypes', describeEndpointTypesResponse_marker - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

$sel:httpStatus:DescribeEndpointTypesResponse', describeEndpointTypesResponse_httpStatus - The response's http status code.

Response Lenses

describeEndpointTypesResponse_supportedEndpointTypes :: Lens' DescribeEndpointTypesResponse (Maybe [SupportedEndpointType])

The types of endpoints that are supported.

describeEndpointTypesResponse_marker :: Lens' DescribeEndpointTypesResponse (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

describeEndpointTypesResponse_httpStatus :: Lens' DescribeEndpointTypesResponse Int

The response's http status code.