Network.AWS.DMS.DescribeOrderableReplicationInstances

Description

Returns information about the replication instance types that can be created in the specified region.

This operation returns paginated results.

Creating a Request

data DescribeOrderableReplicationInstances

See: newDescribeOrderableReplicationInstances smart constructor.

Constructors

DescribeOrderableReplicationInstances'

Fields

maxRecords :: Maybe Int

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.

marker :: Maybe Text

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

newDescribeOrderableReplicationInstances :: DescribeOrderableReplicationInstances

Create a value of DescribeOrderableReplicationInstances with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxRecords:DescribeOrderableReplicationInstances', describeOrderableReplicationInstances_maxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.

$sel:marker:DescribeOrderableReplicationInstances', describeOrderableReplicationInstances_marker - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Request Lenses

describeOrderableReplicationInstances_maxRecords :: Lens' DescribeOrderableReplicationInstances (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.

describeOrderableReplicationInstances_marker :: Lens' DescribeOrderableReplicationInstances (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Destructuring the Response

data DescribeOrderableReplicationInstancesResponse

See: newDescribeOrderableReplicationInstancesResponse smart constructor.

Constructors

DescribeOrderableReplicationInstancesResponse'

Fields

orderableReplicationInstances :: Maybe [OrderableReplicationInstance]

The order-able replication instances available.

marker :: Maybe Text

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

httpStatus :: Int

The response's http status code.

newDescribeOrderableReplicationInstancesResponse

Arguments

:: Int
$sel:httpStatus:DescribeOrderableReplicationInstancesResponse'

-> DescribeOrderableReplicationInstancesResponse

Create a value of DescribeOrderableReplicationInstancesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:orderableReplicationInstances:DescribeOrderableReplicationInstancesResponse', describeOrderableReplicationInstancesResponse_orderableReplicationInstances - The order-able replication instances available.

$sel:marker:DescribeOrderableReplicationInstances', describeOrderableReplicationInstancesResponse_marker - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

$sel:httpStatus:DescribeOrderableReplicationInstancesResponse', describeOrderableReplicationInstancesResponse_httpStatus - The response's http status code.

Response Lenses

describeOrderableReplicationInstancesResponse_orderableReplicationInstances :: Lens' DescribeOrderableReplicationInstancesResponse (Maybe [OrderableReplicationInstance])

The order-able replication instances available.

describeOrderableReplicationInstancesResponse_marker :: Lens' DescribeOrderableReplicationInstancesResponse (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

describeOrderableReplicationInstancesResponse_httpStatus :: Lens' DescribeOrderableReplicationInstancesResponse Int

The response's http status code.