Network.AWS.DMS.DescribePendingMaintenanceActions

Description

For internal use only

Creating a Request

data DescribePendingMaintenanceActions

See: newDescribePendingMaintenanceActions smart constructor.

Constructors

DescribePendingMaintenanceActions'

Fields

filters :: Maybe [Filter]

replicationInstanceArn :: Maybe Text

The Amazon Resource Name (ARN) of the replication instance.

maxRecords :: Maybe Int

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.

marker :: Maybe Text

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

newDescribePendingMaintenanceActions :: DescribePendingMaintenanceActions

Create a value of DescribePendingMaintenanceActions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:DescribePendingMaintenanceActions', describePendingMaintenanceActions_filters -

$sel:replicationInstanceArn:DescribePendingMaintenanceActions', describePendingMaintenanceActions_replicationInstanceArn - The Amazon Resource Name (ARN) of the replication instance.

$sel:maxRecords:DescribePendingMaintenanceActions', describePendingMaintenanceActions_maxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.

$sel:marker:DescribePendingMaintenanceActions', describePendingMaintenanceActions_marker - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Request Lenses

describePendingMaintenanceActions_filters :: Lens' DescribePendingMaintenanceActions (Maybe [Filter])

describePendingMaintenanceActions_replicationInstanceArn :: Lens' DescribePendingMaintenanceActions (Maybe Text)

The Amazon Resource Name (ARN) of the replication instance.

describePendingMaintenanceActions_maxRecords :: Lens' DescribePendingMaintenanceActions (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.

describePendingMaintenanceActions_marker :: Lens' DescribePendingMaintenanceActions (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Destructuring the Response

data DescribePendingMaintenanceActionsResponse

See: newDescribePendingMaintenanceActionsResponse smart constructor.

Constructors

DescribePendingMaintenanceActionsResponse'

Fields

pendingMaintenanceActions :: Maybe [ResourcePendingMaintenanceActions]

The pending maintenance action.

marker :: Maybe Text

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

httpStatus :: Int

The response's http status code.

newDescribePendingMaintenanceActionsResponse

Arguments

:: Int
$sel:httpStatus:DescribePendingMaintenanceActionsResponse'

-> DescribePendingMaintenanceActionsResponse

Create a value of DescribePendingMaintenanceActionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pendingMaintenanceActions:DescribePendingMaintenanceActionsResponse', describePendingMaintenanceActionsResponse_pendingMaintenanceActions - The pending maintenance action.

$sel:marker:DescribePendingMaintenanceActions', describePendingMaintenanceActionsResponse_marker - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

$sel:httpStatus:DescribePendingMaintenanceActionsResponse', describePendingMaintenanceActionsResponse_httpStatus - The response's http status code.

Response Lenses

describePendingMaintenanceActionsResponse_pendingMaintenanceActions :: Lens' DescribePendingMaintenanceActionsResponse (Maybe [ResourcePendingMaintenanceActions])

The pending maintenance action.

describePendingMaintenanceActionsResponse_marker :: Lens' DescribePendingMaintenanceActionsResponse (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

describePendingMaintenanceActionsResponse_httpStatus :: Lens' DescribePendingMaintenanceActionsResponse Int

The response's http status code.