Network.AWS.DMS.DescribeTableStatistics

Description

Returns table statistics on the database migration task, including table name, rows inserted, rows updated, and rows deleted.

Note that the "last updated" column the DMS console only indicates the time that DMS last updated the table statistics record for a table. It does not indicate the time of the last update to the table.

This operation returns paginated results.

Creating a Request

data DescribeTableStatistics

See: newDescribeTableStatistics smart constructor.

Constructors

DescribeTableStatistics'

Fields

filters :: Maybe [Filter]

Filters applied to table statistics.

Valid filter names: schema-name | table-name | table-state

A combination of filters creates an AND condition where each record matches all specified filters.

maxRecords :: Maybe Int

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 500.

marker :: Maybe Text

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

replicationTaskArn :: Text

The Amazon Resource Name (ARN) of the replication task.

newDescribeTableStatistics

Arguments

:: Text
$sel:replicationTaskArn:DescribeTableStatistics'

-> DescribeTableStatistics

Create a value of DescribeTableStatistics with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:DescribeTableStatistics', describeTableStatistics_filters - Filters applied to table statistics.

Valid filter names: schema-name | table-name | table-state

A combination of filters creates an AND condition where each record matches all specified filters.

$sel:maxRecords:DescribeTableStatistics', describeTableStatistics_maxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 500.

$sel:marker:DescribeTableStatistics', describeTableStatistics_marker - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

$sel:replicationTaskArn:DescribeTableStatistics', describeTableStatistics_replicationTaskArn - The Amazon Resource Name (ARN) of the replication task.

Request Lenses

describeTableStatistics_filters :: Lens' DescribeTableStatistics (Maybe [Filter])

Filters applied to table statistics.

Valid filter names: schema-name | table-name | table-state

A combination of filters creates an AND condition where each record matches all specified filters.

describeTableStatistics_maxRecords :: Lens' DescribeTableStatistics (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 500.

describeTableStatistics_marker :: Lens' DescribeTableStatistics (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

describeTableStatistics_replicationTaskArn :: Lens' DescribeTableStatistics Text

The Amazon Resource Name (ARN) of the replication task.

Destructuring the Response

data DescribeTableStatisticsResponse

See: newDescribeTableStatisticsResponse smart constructor.

Constructors

DescribeTableStatisticsResponse'

Fields

tableStatistics :: Maybe [TableStatistics]

The table statistics.

replicationTaskArn :: Maybe Text

The Amazon Resource Name (ARN) of the replication task.

marker :: Maybe Text

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

httpStatus :: Int

The response's http status code.

newDescribeTableStatisticsResponse

Arguments

:: Int
$sel:httpStatus:DescribeTableStatisticsResponse'

-> DescribeTableStatisticsResponse

Create a value of DescribeTableStatisticsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tableStatistics:DescribeTableStatisticsResponse', describeTableStatisticsResponse_tableStatistics - The table statistics.

$sel:replicationTaskArn:DescribeTableStatistics', describeTableStatisticsResponse_replicationTaskArn - The Amazon Resource Name (ARN) of the replication task.

$sel:marker:DescribeTableStatistics', describeTableStatisticsResponse_marker - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

$sel:httpStatus:DescribeTableStatisticsResponse', describeTableStatisticsResponse_httpStatus - The response's http status code.

Response Lenses

describeTableStatisticsResponse_tableStatistics :: Lens' DescribeTableStatisticsResponse (Maybe [TableStatistics])

The table statistics.

describeTableStatisticsResponse_replicationTaskArn :: Lens' DescribeTableStatisticsResponse (Maybe Text)

The Amazon Resource Name (ARN) of the replication task.

describeTableStatisticsResponse_marker :: Lens' DescribeTableStatisticsResponse (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

describeTableStatisticsResponse_httpStatus :: Lens' DescribeTableStatisticsResponse Int

The response's http status code.