Network.AWS.DMS.ImportCertificate

Description

Uploads the specified certificate.

Creating a Request

data ImportCertificate

See: newImportCertificate smart constructor.

Constructors

ImportCertificate'

Fields

certificatePem :: Maybe (Sensitive Text)

The contents of a .pem file, which contains an X.509 certificate.

certificateWallet :: Maybe Base64

The location of an imported Oracle Wallet certificate for use with SSL. Provide the name of a .sso file using the fileb:// prefix. You can't provide the certificate inline.

tags :: Maybe [Tag]

The tags associated with the certificate.

certificateIdentifier :: Text

A customer-assigned name for the certificate. Identifiers must begin with a letter and must contain only ASCII letters, digits, and hyphens. They can't end with a hyphen or contain two consecutive hyphens.

newImportCertificate

Arguments

:: Text
$sel:certificateIdentifier:ImportCertificate'

-> ImportCertificate

Create a value of ImportCertificate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:certificatePem:ImportCertificate', importCertificate_certificatePem - The contents of a .pem file, which contains an X.509 certificate.

$sel:certificateWallet:ImportCertificate', importCertificate_certificateWallet - The location of an imported Oracle Wallet certificate for use with SSL. Provide the name of a .sso file using the fileb:// prefix. You can't provide the certificate inline.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:tags:ImportCertificate', importCertificate_tags - The tags associated with the certificate.

$sel:certificateIdentifier:ImportCertificate', importCertificate_certificateIdentifier - A customer-assigned name for the certificate. Identifiers must begin with a letter and must contain only ASCII letters, digits, and hyphens. They can't end with a hyphen or contain two consecutive hyphens.

Request Lenses

importCertificate_certificatePem :: Lens' ImportCertificate (Maybe Text)

The contents of a .pem file, which contains an X.509 certificate.

importCertificate_certificateWallet :: Lens' ImportCertificate (Maybe ByteString)

The location of an imported Oracle Wallet certificate for use with SSL. Provide the name of a .sso file using the fileb:// prefix. You can't provide the certificate inline.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

importCertificate_tags :: Lens' ImportCertificate (Maybe [Tag])

The tags associated with the certificate.

importCertificate_certificateIdentifier :: Lens' ImportCertificate Text

A customer-assigned name for the certificate. Identifiers must begin with a letter and must contain only ASCII letters, digits, and hyphens. They can't end with a hyphen or contain two consecutive hyphens.

Destructuring the Response

data ImportCertificateResponse

See: newImportCertificateResponse smart constructor.

Constructors

ImportCertificateResponse'

Fields

certificate :: Maybe Certificate

The certificate to be uploaded.

httpStatus :: Int

The response's http status code.

newImportCertificateResponse

Arguments

:: Int
$sel:httpStatus:ImportCertificateResponse'

-> ImportCertificateResponse

Create a value of ImportCertificateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:certificate:ImportCertificateResponse', importCertificateResponse_certificate - The certificate to be uploaded.

$sel:httpStatus:ImportCertificateResponse', importCertificateResponse_httpStatus - The response's http status code.

Response Lenses

importCertificateResponse_certificate :: Lens' ImportCertificateResponse (Maybe Certificate)

The certificate to be uploaded.

importCertificateResponse_httpStatus :: Lens' ImportCertificateResponse Int

The response's http status code.