Network.AWS.DMS.ModifyEndpoint

Description

Modifies the specified endpoint.

For a MySQL source or target endpoint, don't explicitly specify the database using the DatabaseName request parameter on the ModifyEndpoint API call. Specifying DatabaseName when you modify a MySQL endpoint replicates all the task tables to this single database. For MySQL endpoints, you specify the database only when you specify the schema in the table-mapping rules of the DMS task.

Creating a Request

data ModifyEndpoint

See: newModifyEndpoint smart constructor.

Constructors

ModifyEndpoint'

Fields

mongoDbSettings :: Maybe MongoDbSettings

Settings in JSON format for the source MongoDB endpoint. For more information about the available settings, see the configuration properties section in Endpoint configuration settings when using MongoDB as a source for Database Migration Service in the Database Migration Service User Guide.

sslMode :: Maybe DmsSslModeValue

The SSL mode used to connect to the endpoint. The default value is none.

neptuneSettings :: Maybe NeptuneSettings

Settings in JSON format for the target Amazon Neptune endpoint. For more information about the available settings, see Specifying graph-mapping rules using Gremlin and R2RML for Amazon Neptune as a target in the Database Migration Service User Guide.

redisSettings :: Maybe RedisSettings

Settings in JSON format for the Redis target endpoint.

engineName :: Maybe Text

The type of engine for the endpoint. Valid values, depending on the EndpointType, include "mysql", "oracle", "postgres", "mariadb", "aurora", "aurora-postgresql", "redshift", "s3", "db2", "azuredb", "sybase", "dynamodb", "mongodb", "kinesis", "kafka", "elasticsearch", "documentdb", "sqlserver", and "neptune".

exactSettings :: Maybe Bool

If this attribute is Y, the current call to ModifyEndpoint replaces all existing endpoint settings with the exact settings that you specify in this call. If this attribute is N, the current call to ModifyEndpoint does two things:

It replaces any endpoint settings that already exist with new values, for settings with the same names.

It creates new endpoint settings that you specify in the call, for settings with different names.

For example, if you call create-endpoint ... --endpoint-settings '{"a":1}' ..., the endpoint has the following endpoint settings: '{"a":1}'. If you then call modify-endpoint ... --endpoint-settings '{"b":2}' ... for the same endpoint, the endpoint has the following settings: '{"a":1,"b":2}'.

However, suppose that you follow this with a call to modify-endpoint ... --endpoint-settings '{"b":2}' --exact-settings ... for that same endpoint again. Then the endpoint has the following settings: '{"b":2}'. All existing settings are replaced with the exact settings that you specify.

externalTableDefinition :: Maybe Text

The external table definition.

elasticsearchSettings :: Maybe ElasticsearchSettings

Settings in JSON format for the target Elasticsearch endpoint. For more information about the available settings, see Extra Connection Attributes When Using Elasticsearch as a Target for DMS in the Database Migration Service User Guide.

postgreSQLSettings :: Maybe PostgreSQLSettings

Settings in JSON format for the source and target PostgreSQL endpoint. For information about other available settings, see Extra connection attributes when using PostgreSQL as a source for DMS and Extra connection attributes when using PostgreSQL as a target for DMS in the Database Migration Service User Guide.

oracleSettings :: Maybe OracleSettings

Settings in JSON format for the source and target Oracle endpoint. For information about other available settings, see Extra connection attributes when using Oracle as a source for DMS and Extra connection attributes when using Oracle as a target for DMS in the Database Migration Service User Guide.

endpointType :: Maybe ReplicationEndpointTypeValue

The type of endpoint. Valid values are source and target.

certificateArn :: Maybe Text

The Amazon Resource Name (ARN) of the certificate used for SSL connection.

serviceAccessRoleArn :: Maybe Text

The Amazon Resource Name (ARN) for the IAM role you want to use to modify the endpoint. The role must allow the iam:PassRole action.

s3Settings :: Maybe S3Settings

Settings in JSON format for the target Amazon S3 endpoint. For more information about the available settings, see Extra Connection Attributes When Using Amazon S3 as a Target for DMS in the Database Migration Service User Guide.

microsoftSQLServerSettings :: Maybe MicrosoftSQLServerSettings

Settings in JSON format for the source and target Microsoft SQL Server endpoint. For information about other available settings, see Extra connection attributes when using SQL Server as a source for DMS and Extra connection attributes when using SQL Server as a target for DMS in the Database Migration Service User Guide.

serverName :: Maybe Text

The name of the server where the endpoint database resides.

iBMDb2Settings :: Maybe IBMDb2Settings

Settings in JSON format for the source IBM Db2 LUW endpoint. For information about other available settings, see Extra connection attributes when using Db2 LUW as a source for DMS in the Database Migration Service User Guide.

password :: Maybe (Sensitive Text)

The password to be used to login to the endpoint database.

mySQLSettings :: Maybe MySQLSettings

Settings in JSON format for the source and target MySQL endpoint. For information about other available settings, see Extra connection attributes when using MySQL as a source for DMS and Extra connection attributes when using a MySQL-compatible database as a target for DMS in the Database Migration Service User Guide.

dmsTransferSettings :: Maybe DmsTransferSettings

The settings in JSON format for the DMS transfer type of source endpoint.

Attributes include the following:

serviceAccessRoleArn - The Amazon Resource Name (ARN) used by the service access IAM role. The role must allow the iam:PassRole action.

BucketName - The name of the S3 bucket to use.

Shorthand syntax for these settings is as follows: ServiceAccessRoleArn=string ,BucketName=string

JSON syntax for these settings is as follows: { "ServiceAccessRoleArn": "string", "BucketName": "string"}

port :: Maybe Int

The port used by the endpoint database.

redshiftSettings :: Maybe RedshiftSettings

username :: Maybe Text

The user name to be used to login to the endpoint database.

extraConnectionAttributes :: Maybe Text

Additional attributes associated with the connection. To reset this parameter, pass the empty string ("") as an argument.

kafkaSettings :: Maybe KafkaSettings

Settings in JSON format for the target Apache Kafka endpoint. For more information about the available settings, see Using object mapping to migrate data to a Kafka topic in the Database Migration Service User Guide.

docDbSettings :: Maybe DocDbSettings

Settings in JSON format for the source DocumentDB endpoint. For more information about the available settings, see the configuration properties section in Using DocumentDB as a Target for Database Migration Service in the Database Migration Service User Guide.

dynamoDbSettings :: Maybe DynamoDbSettings

Settings in JSON format for the target Amazon DynamoDB endpoint. For information about other available settings, see Using Object Mapping to Migrate Data to DynamoDB in the Database Migration Service User Guide.

endpointIdentifier :: Maybe Text

The database endpoint identifier. Identifiers must begin with a letter and must contain only ASCII letters, digits, and hyphens. They can't end with a hyphen or contain two consecutive hyphens.

kinesisSettings :: Maybe KinesisSettings

Settings in JSON format for the target endpoint for Amazon Kinesis Data Streams. For more information about the available settings, see Using object mapping to migrate data to a Kinesis data stream in the Database Migration Service User Guide.

sybaseSettings :: Maybe SybaseSettings

Settings in JSON format for the source and target SAP ASE endpoint. For information about other available settings, see Extra connection attributes when using SAP ASE as a source for DMS and Extra connection attributes when using SAP ASE as a target for DMS in the Database Migration Service User Guide.

databaseName :: Maybe Text

The name of the endpoint database. For a MySQL source or target endpoint, do not specify DatabaseName.

endpointArn :: Text

The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.

newModifyEndpoint

Arguments

:: Text
$sel:endpointArn:ModifyEndpoint'

-> ModifyEndpoint

Create a value of ModifyEndpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:mongoDbSettings:ModifyEndpoint', modifyEndpoint_mongoDbSettings - Settings in JSON format for the source MongoDB endpoint. For more information about the available settings, see the configuration properties section in Endpoint configuration settings when using MongoDB as a source for Database Migration Service in the Database Migration Service User Guide.

$sel:sslMode:ModifyEndpoint', modifyEndpoint_sslMode - The SSL mode used to connect to the endpoint. The default value is none.

$sel:neptuneSettings:ModifyEndpoint', modifyEndpoint_neptuneSettings - Settings in JSON format for the target Amazon Neptune endpoint. For more information about the available settings, see Specifying graph-mapping rules using Gremlin and R2RML for Amazon Neptune as a target in the Database Migration Service User Guide.

$sel:redisSettings:ModifyEndpoint', modifyEndpoint_redisSettings - Settings in JSON format for the Redis target endpoint.

$sel:engineName:ModifyEndpoint', modifyEndpoint_engineName - The type of engine for the endpoint. Valid values, depending on the EndpointType, include "mysql", "oracle", "postgres", "mariadb", "aurora", "aurora-postgresql", "redshift", "s3", "db2", "azuredb", "sybase", "dynamodb", "mongodb", "kinesis", "kafka", "elasticsearch", "documentdb", "sqlserver", and "neptune".

$sel:exactSettings:ModifyEndpoint', modifyEndpoint_exactSettings - If this attribute is Y, the current call to ModifyEndpoint replaces all existing endpoint settings with the exact settings that you specify in this call. If this attribute is N, the current call to ModifyEndpoint does two things:

It replaces any endpoint settings that already exist with new values, for settings with the same names.

It creates new endpoint settings that you specify in the call, for settings with different names.

For example, if you call create-endpoint ... --endpoint-settings '{"a":1}' ..., the endpoint has the following endpoint settings: '{"a":1}'. If you then call modify-endpoint ... --endpoint-settings '{"b":2}' ... for the same endpoint, the endpoint has the following settings: '{"a":1,"b":2}'.

However, suppose that you follow this with a call to modify-endpoint ... --endpoint-settings '{"b":2}' --exact-settings ... for that same endpoint again. Then the endpoint has the following settings: '{"b":2}'. All existing settings are replaced with the exact settings that you specify.

$sel:externalTableDefinition:ModifyEndpoint', modifyEndpoint_externalTableDefinition - The external table definition.

$sel:elasticsearchSettings:ModifyEndpoint', modifyEndpoint_elasticsearchSettings - Settings in JSON format for the target Elasticsearch endpoint. For more information about the available settings, see Extra Connection Attributes When Using Elasticsearch as a Target for DMS in the Database Migration Service User Guide.

$sel:postgreSQLSettings:ModifyEndpoint', modifyEndpoint_postgreSQLSettings - Settings in JSON format for the source and target PostgreSQL endpoint. For information about other available settings, see Extra connection attributes when using PostgreSQL as a source for DMS and Extra connection attributes when using PostgreSQL as a target for DMS in the Database Migration Service User Guide.

$sel:oracleSettings:ModifyEndpoint', modifyEndpoint_oracleSettings - Settings in JSON format for the source and target Oracle endpoint. For information about other available settings, see Extra connection attributes when using Oracle as a source for DMS and Extra connection attributes when using Oracle as a target for DMS in the Database Migration Service User Guide.

$sel:endpointType:ModifyEndpoint', modifyEndpoint_endpointType - The type of endpoint. Valid values are source and target.

$sel:certificateArn:ModifyEndpoint', modifyEndpoint_certificateArn - The Amazon Resource Name (ARN) of the certificate used for SSL connection.

$sel:serviceAccessRoleArn:ModifyEndpoint', modifyEndpoint_serviceAccessRoleArn - The Amazon Resource Name (ARN) for the IAM role you want to use to modify the endpoint. The role must allow the iam:PassRole action.

$sel:s3Settings:ModifyEndpoint', modifyEndpoint_s3Settings - Settings in JSON format for the target Amazon S3 endpoint. For more information about the available settings, see Extra Connection Attributes When Using Amazon S3 as a Target for DMS in the Database Migration Service User Guide.

$sel:microsoftSQLServerSettings:ModifyEndpoint', modifyEndpoint_microsoftSQLServerSettings - Settings in JSON format for the source and target Microsoft SQL Server endpoint. For information about other available settings, see Extra connection attributes when using SQL Server as a source for DMS and Extra connection attributes when using SQL Server as a target for DMS in the Database Migration Service User Guide.

$sel:serverName:ModifyEndpoint', modifyEndpoint_serverName - The name of the server where the endpoint database resides.

$sel:iBMDb2Settings:ModifyEndpoint', modifyEndpoint_iBMDb2Settings - Settings in JSON format for the source IBM Db2 LUW endpoint. For information about other available settings, see Extra connection attributes when using Db2 LUW as a source for DMS in the Database Migration Service User Guide.

$sel:password:ModifyEndpoint', modifyEndpoint_password - The password to be used to login to the endpoint database.

$sel:mySQLSettings:ModifyEndpoint', modifyEndpoint_mySQLSettings - Settings in JSON format for the source and target MySQL endpoint. For information about other available settings, see Extra connection attributes when using MySQL as a source for DMS and Extra connection attributes when using a MySQL-compatible database as a target for DMS in the Database Migration Service User Guide.

$sel:dmsTransferSettings:ModifyEndpoint', modifyEndpoint_dmsTransferSettings - The settings in JSON format for the DMS transfer type of source endpoint.

Attributes include the following:

serviceAccessRoleArn - The Amazon Resource Name (ARN) used by the service access IAM role. The role must allow the iam:PassRole action.

BucketName - The name of the S3 bucket to use.

Shorthand syntax for these settings is as follows: ServiceAccessRoleArn=string ,BucketName=string

JSON syntax for these settings is as follows: { "ServiceAccessRoleArn": "string", "BucketName": "string"}

$sel:port:ModifyEndpoint', modifyEndpoint_port - The port used by the endpoint database.

$sel:redshiftSettings:ModifyEndpoint', modifyEndpoint_redshiftSettings - Undocumented member.

$sel:username:ModifyEndpoint', modifyEndpoint_username - The user name to be used to login to the endpoint database.

$sel:extraConnectionAttributes:ModifyEndpoint', modifyEndpoint_extraConnectionAttributes - Additional attributes associated with the connection. To reset this parameter, pass the empty string ("") as an argument.

$sel:kafkaSettings:ModifyEndpoint', modifyEndpoint_kafkaSettings - Settings in JSON format for the target Apache Kafka endpoint. For more information about the available settings, see Using object mapping to migrate data to a Kafka topic in the Database Migration Service User Guide.

$sel:docDbSettings:ModifyEndpoint', modifyEndpoint_docDbSettings - Settings in JSON format for the source DocumentDB endpoint. For more information about the available settings, see the configuration properties section in Using DocumentDB as a Target for Database Migration Service in the Database Migration Service User Guide.

$sel:dynamoDbSettings:ModifyEndpoint', modifyEndpoint_dynamoDbSettings - Settings in JSON format for the target Amazon DynamoDB endpoint. For information about other available settings, see Using Object Mapping to Migrate Data to DynamoDB in the Database Migration Service User Guide.

$sel:endpointIdentifier:ModifyEndpoint', modifyEndpoint_endpointIdentifier - The database endpoint identifier. Identifiers must begin with a letter and must contain only ASCII letters, digits, and hyphens. They can't end with a hyphen or contain two consecutive hyphens.

$sel:kinesisSettings:ModifyEndpoint', modifyEndpoint_kinesisSettings - Settings in JSON format for the target endpoint for Amazon Kinesis Data Streams. For more information about the available settings, see Using object mapping to migrate data to a Kinesis data stream in the Database Migration Service User Guide.

$sel:sybaseSettings:ModifyEndpoint', modifyEndpoint_sybaseSettings - Settings in JSON format for the source and target SAP ASE endpoint. For information about other available settings, see Extra connection attributes when using SAP ASE as a source for DMS and Extra connection attributes when using SAP ASE as a target for DMS in the Database Migration Service User Guide.

$sel:databaseName:ModifyEndpoint', modifyEndpoint_databaseName - The name of the endpoint database. For a MySQL source or target endpoint, do not specify DatabaseName.

$sel:endpointArn:ModifyEndpoint', modifyEndpoint_endpointArn - The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.

Request Lenses

modifyEndpoint_mongoDbSettings :: Lens' ModifyEndpoint (Maybe MongoDbSettings)

Settings in JSON format for the source MongoDB endpoint. For more information about the available settings, see the configuration properties section in Endpoint configuration settings when using MongoDB as a source for Database Migration Service in the Database Migration Service User Guide.

modifyEndpoint_sslMode :: Lens' ModifyEndpoint (Maybe DmsSslModeValue)

The SSL mode used to connect to the endpoint. The default value is none.

modifyEndpoint_neptuneSettings :: Lens' ModifyEndpoint (Maybe NeptuneSettings)

Settings in JSON format for the target Amazon Neptune endpoint. For more information about the available settings, see Specifying graph-mapping rules using Gremlin and R2RML for Amazon Neptune as a target in the Database Migration Service User Guide.

modifyEndpoint_redisSettings :: Lens' ModifyEndpoint (Maybe RedisSettings)

Settings in JSON format for the Redis target endpoint.

modifyEndpoint_engineName :: Lens' ModifyEndpoint (Maybe Text)

The type of engine for the endpoint. Valid values, depending on the EndpointType, include "mysql", "oracle", "postgres", "mariadb", "aurora", "aurora-postgresql", "redshift", "s3", "db2", "azuredb", "sybase", "dynamodb", "mongodb", "kinesis", "kafka", "elasticsearch", "documentdb", "sqlserver", and "neptune".

modifyEndpoint_exactSettings :: Lens' ModifyEndpoint (Maybe Bool)

If this attribute is Y, the current call to ModifyEndpoint replaces all existing endpoint settings with the exact settings that you specify in this call. If this attribute is N, the current call to ModifyEndpoint does two things:

It replaces any endpoint settings that already exist with new values, for settings with the same names.

It creates new endpoint settings that you specify in the call, for settings with different names.

For example, if you call create-endpoint ... --endpoint-settings '{"a":1}' ..., the endpoint has the following endpoint settings: '{"a":1}'. If you then call modify-endpoint ... --endpoint-settings '{"b":2}' ... for the same endpoint, the endpoint has the following settings: '{"a":1,"b":2}'.

However, suppose that you follow this with a call to modify-endpoint ... --endpoint-settings '{"b":2}' --exact-settings ... for that same endpoint again. Then the endpoint has the following settings: '{"b":2}'. All existing settings are replaced with the exact settings that you specify.

modifyEndpoint_externalTableDefinition :: Lens' ModifyEndpoint (Maybe Text)

The external table definition.

modifyEndpoint_elasticsearchSettings :: Lens' ModifyEndpoint (Maybe ElasticsearchSettings)

Settings in JSON format for the target Elasticsearch endpoint. For more information about the available settings, see Extra Connection Attributes When Using Elasticsearch as a Target for DMS in the Database Migration Service User Guide.

modifyEndpoint_postgreSQLSettings :: Lens' ModifyEndpoint (Maybe PostgreSQLSettings)

Settings in JSON format for the source and target PostgreSQL endpoint. For information about other available settings, see Extra connection attributes when using PostgreSQL as a source for DMS and Extra connection attributes when using PostgreSQL as a target for DMS in the Database Migration Service User Guide.

modifyEndpoint_oracleSettings :: Lens' ModifyEndpoint (Maybe OracleSettings)

Settings in JSON format for the source and target Oracle endpoint. For information about other available settings, see Extra connection attributes when using Oracle as a source for DMS and Extra connection attributes when using Oracle as a target for DMS in the Database Migration Service User Guide.

modifyEndpoint_endpointType :: Lens' ModifyEndpoint (Maybe ReplicationEndpointTypeValue)

The type of endpoint. Valid values are source and target.

modifyEndpoint_certificateArn :: Lens' ModifyEndpoint (Maybe Text)

The Amazon Resource Name (ARN) of the certificate used for SSL connection.

modifyEndpoint_serviceAccessRoleArn :: Lens' ModifyEndpoint (Maybe Text)

The Amazon Resource Name (ARN) for the IAM role you want to use to modify the endpoint. The role must allow the iam:PassRole action.

modifyEndpoint_s3Settings :: Lens' ModifyEndpoint (Maybe S3Settings)

Settings in JSON format for the target Amazon S3 endpoint. For more information about the available settings, see Extra Connection Attributes When Using Amazon S3 as a Target for DMS in the Database Migration Service User Guide.

modifyEndpoint_microsoftSQLServerSettings :: Lens' ModifyEndpoint (Maybe MicrosoftSQLServerSettings)

Settings in JSON format for the source and target Microsoft SQL Server endpoint. For information about other available settings, see Extra connection attributes when using SQL Server as a source for DMS and Extra connection attributes when using SQL Server as a target for DMS in the Database Migration Service User Guide.

modifyEndpoint_serverName :: Lens' ModifyEndpoint (Maybe Text)

The name of the server where the endpoint database resides.

modifyEndpoint_iBMDb2Settings :: Lens' ModifyEndpoint (Maybe IBMDb2Settings)

Settings in JSON format for the source IBM Db2 LUW endpoint. For information about other available settings, see Extra connection attributes when using Db2 LUW as a source for DMS in the Database Migration Service User Guide.

modifyEndpoint_password :: Lens' ModifyEndpoint (Maybe Text)

The password to be used to login to the endpoint database.

modifyEndpoint_mySQLSettings :: Lens' ModifyEndpoint (Maybe MySQLSettings)

Settings in JSON format for the source and target MySQL endpoint. For information about other available settings, see Extra connection attributes when using MySQL as a source for DMS and Extra connection attributes when using a MySQL-compatible database as a target for DMS in the Database Migration Service User Guide.

modifyEndpoint_dmsTransferSettings :: Lens' ModifyEndpoint (Maybe DmsTransferSettings)

The settings in JSON format for the DMS transfer type of source endpoint.

Attributes include the following:

serviceAccessRoleArn - The Amazon Resource Name (ARN) used by the service access IAM role. The role must allow the iam:PassRole action.

BucketName - The name of the S3 bucket to use.

Shorthand syntax for these settings is as follows: ServiceAccessRoleArn=string ,BucketName=string

JSON syntax for these settings is as follows: { "ServiceAccessRoleArn": "string", "BucketName": "string"}

modifyEndpoint_port :: Lens' ModifyEndpoint (Maybe Int)

The port used by the endpoint database.

modifyEndpoint_redshiftSettings :: Lens' ModifyEndpoint (Maybe RedshiftSettings)

Undocumented member.

modifyEndpoint_username :: Lens' ModifyEndpoint (Maybe Text)

The user name to be used to login to the endpoint database.

modifyEndpoint_extraConnectionAttributes :: Lens' ModifyEndpoint (Maybe Text)

Additional attributes associated with the connection. To reset this parameter, pass the empty string ("") as an argument.

modifyEndpoint_kafkaSettings :: Lens' ModifyEndpoint (Maybe KafkaSettings)

Settings in JSON format for the target Apache Kafka endpoint. For more information about the available settings, see Using object mapping to migrate data to a Kafka topic in the Database Migration Service User Guide.

modifyEndpoint_docDbSettings :: Lens' ModifyEndpoint (Maybe DocDbSettings)

Settings in JSON format for the source DocumentDB endpoint. For more information about the available settings, see the configuration properties section in Using DocumentDB as a Target for Database Migration Service in the Database Migration Service User Guide.

modifyEndpoint_dynamoDbSettings :: Lens' ModifyEndpoint (Maybe DynamoDbSettings)

Settings in JSON format for the target Amazon DynamoDB endpoint. For information about other available settings, see Using Object Mapping to Migrate Data to DynamoDB in the Database Migration Service User Guide.

modifyEndpoint_endpointIdentifier :: Lens' ModifyEndpoint (Maybe Text)

The database endpoint identifier. Identifiers must begin with a letter and must contain only ASCII letters, digits, and hyphens. They can't end with a hyphen or contain two consecutive hyphens.

modifyEndpoint_kinesisSettings :: Lens' ModifyEndpoint (Maybe KinesisSettings)

Settings in JSON format for the target endpoint for Amazon Kinesis Data Streams. For more information about the available settings, see Using object mapping to migrate data to a Kinesis data stream in the Database Migration Service User Guide.

modifyEndpoint_sybaseSettings :: Lens' ModifyEndpoint (Maybe SybaseSettings)

Settings in JSON format for the source and target SAP ASE endpoint. For information about other available settings, see Extra connection attributes when using SAP ASE as a source for DMS and Extra connection attributes when using SAP ASE as a target for DMS in the Database Migration Service User Guide.

modifyEndpoint_databaseName :: Lens' ModifyEndpoint (Maybe Text)

The name of the endpoint database. For a MySQL source or target endpoint, do not specify DatabaseName.

modifyEndpoint_endpointArn :: Lens' ModifyEndpoint Text

The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.

Destructuring the Response

data ModifyEndpointResponse

See: newModifyEndpointResponse smart constructor.

Constructors

ModifyEndpointResponse'

Fields

endpoint :: Maybe Endpoint

The modified endpoint.

httpStatus :: Int

The response's http status code.

newModifyEndpointResponse

Arguments

:: Int
$sel:httpStatus:ModifyEndpointResponse'

-> ModifyEndpointResponse

Create a value of ModifyEndpointResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:endpoint:ModifyEndpointResponse', modifyEndpointResponse_endpoint - The modified endpoint.

$sel:httpStatus:ModifyEndpointResponse', modifyEndpointResponse_httpStatus - The response's http status code.

Response Lenses

modifyEndpointResponse_endpoint :: Lens' ModifyEndpointResponse (Maybe Endpoint)

The modified endpoint.

modifyEndpointResponse_httpStatus :: Lens' ModifyEndpointResponse Int

The response's http status code.