Network.AWS.DMS.ModifyReplicationInstance

Description

Modifies the replication instance to apply new settings. You can change one or more parameters by specifying these parameters and the new values in the request.

Some settings are applied during the maintenance window.

Creating a Request

data ModifyReplicationInstance

See: newModifyReplicationInstance smart constructor.

Constructors

ModifyReplicationInstance'

Fields

allowMajorVersionUpgrade :: Maybe Bool

Indicates that major version upgrades are allowed. Changing this parameter does not result in an outage, and the change is asynchronously applied as soon as possible.

This parameter must be set to true when specifying a value for the EngineVersion parameter that is a different major version than the replication instance's current version.

multiAZ :: Maybe Bool

Specifies whether the replication instance is a Multi-AZ deployment. You can't set the AvailabilityZone parameter if the Multi-AZ parameter is set to true.

vpcSecurityGroupIds :: Maybe [Text]

Specifies the VPC security group to be used with the replication instance. The VPC security group must work with the VPC containing the replication instance.

engineVersion :: Maybe Text

The engine version number of the replication instance.

When modifying a major engine version of an instance, also set AllowMajorVersionUpgrade to true.

preferredMaintenanceWindow :: Maybe Text

The weekly time range (in UTC) during which system maintenance can occur, which might result in an outage. Changing this parameter does not result in an outage, except in the following situation, and the change is asynchronously applied as soon as possible. If moving this window to the current time, there must be at least 30 minutes between the current time and end of the window to ensure pending changes are applied.

Default: Uses existing setting

Format: ddd:hh24:mi-ddd:hh24:mi

Valid Days: Mon | Tue | Wed | Thu | Fri | Sat | Sun

Constraints: Must be at least 30 minutes

replicationInstanceIdentifier :: Maybe Text

The replication instance identifier. This parameter is stored as a lowercase string.

replicationInstanceClass :: Maybe Text

The compute and memory capacity of the replication instance as defined for the specified replication instance class. For example to specify the instance class dms.c4.large, set this parameter to "dms.c4.large".

For more information on the settings and capacities for the available replication instance classes, see Selecting the right DMS replication instance for your migration.

allocatedStorage :: Maybe Int

The amount of storage (in gigabytes) to be allocated for the replication instance.

applyImmediately :: Maybe Bool

Indicates whether the changes should be applied immediately or during the next maintenance window.

autoMinorVersionUpgrade :: Maybe Bool

A value that indicates that minor version upgrades are applied automatically to the replication instance during the maintenance window. Changing this parameter doesn't result in an outage, except in the case described following. The change is asynchronously applied as soon as possible.

An outage does result if these factors apply:

This parameter is set to true during the maintenance window.

A newer minor version is available.

DMS has enabled automatic patching for the given engine version.

replicationInstanceArn :: Text

The Amazon Resource Name (ARN) of the replication instance.

newModifyReplicationInstance

Arguments

:: Text
$sel:replicationInstanceArn:ModifyReplicationInstance'

-> ModifyReplicationInstance

Create a value of ModifyReplicationInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:allowMajorVersionUpgrade:ModifyReplicationInstance', modifyReplicationInstance_allowMajorVersionUpgrade - Indicates that major version upgrades are allowed. Changing this parameter does not result in an outage, and the change is asynchronously applied as soon as possible.

This parameter must be set to true when specifying a value for the EngineVersion parameter that is a different major version than the replication instance's current version.

$sel:multiAZ:ModifyReplicationInstance', modifyReplicationInstance_multiAZ - Specifies whether the replication instance is a Multi-AZ deployment. You can't set the AvailabilityZone parameter if the Multi-AZ parameter is set to true.

$sel:vpcSecurityGroupIds:ModifyReplicationInstance', modifyReplicationInstance_vpcSecurityGroupIds - Specifies the VPC security group to be used with the replication instance. The VPC security group must work with the VPC containing the replication instance.

$sel:engineVersion:ModifyReplicationInstance', modifyReplicationInstance_engineVersion - The engine version number of the replication instance.

When modifying a major engine version of an instance, also set AllowMajorVersionUpgrade to true.

$sel:preferredMaintenanceWindow:ModifyReplicationInstance', modifyReplicationInstance_preferredMaintenanceWindow - The weekly time range (in UTC) during which system maintenance can occur, which might result in an outage. Changing this parameter does not result in an outage, except in the following situation, and the change is asynchronously applied as soon as possible. If moving this window to the current time, there must be at least 30 minutes between the current time and end of the window to ensure pending changes are applied.

Default: Uses existing setting

Format: ddd:hh24:mi-ddd:hh24:mi

Valid Days: Mon | Tue | Wed | Thu | Fri | Sat | Sun

Constraints: Must be at least 30 minutes

$sel:replicationInstanceIdentifier:ModifyReplicationInstance', modifyReplicationInstance_replicationInstanceIdentifier - The replication instance identifier. This parameter is stored as a lowercase string.

$sel:replicationInstanceClass:ModifyReplicationInstance', modifyReplicationInstance_replicationInstanceClass - The compute and memory capacity of the replication instance as defined for the specified replication instance class. For example to specify the instance class dms.c4.large, set this parameter to "dms.c4.large".

For more information on the settings and capacities for the available replication instance classes, see Selecting the right DMS replication instance for your migration.

$sel:allocatedStorage:ModifyReplicationInstance', modifyReplicationInstance_allocatedStorage - The amount of storage (in gigabytes) to be allocated for the replication instance.

$sel:applyImmediately:ModifyReplicationInstance', modifyReplicationInstance_applyImmediately - Indicates whether the changes should be applied immediately or during the next maintenance window.

$sel:autoMinorVersionUpgrade:ModifyReplicationInstance', modifyReplicationInstance_autoMinorVersionUpgrade - A value that indicates that minor version upgrades are applied automatically to the replication instance during the maintenance window. Changing this parameter doesn't result in an outage, except in the case described following. The change is asynchronously applied as soon as possible.

An outage does result if these factors apply:

This parameter is set to true during the maintenance window.

A newer minor version is available.

DMS has enabled automatic patching for the given engine version.

$sel:replicationInstanceArn:ModifyReplicationInstance', modifyReplicationInstance_replicationInstanceArn - The Amazon Resource Name (ARN) of the replication instance.

Request Lenses

modifyReplicationInstance_allowMajorVersionUpgrade :: Lens' ModifyReplicationInstance (Maybe Bool)

Indicates that major version upgrades are allowed. Changing this parameter does not result in an outage, and the change is asynchronously applied as soon as possible.

This parameter must be set to true when specifying a value for the EngineVersion parameter that is a different major version than the replication instance's current version.

modifyReplicationInstance_multiAZ :: Lens' ModifyReplicationInstance (Maybe Bool)

Specifies whether the replication instance is a Multi-AZ deployment. You can't set the AvailabilityZone parameter if the Multi-AZ parameter is set to true.

modifyReplicationInstance_vpcSecurityGroupIds :: Lens' ModifyReplicationInstance (Maybe [Text])

Specifies the VPC security group to be used with the replication instance. The VPC security group must work with the VPC containing the replication instance.

modifyReplicationInstance_engineVersion :: Lens' ModifyReplicationInstance (Maybe Text)

The engine version number of the replication instance.

When modifying a major engine version of an instance, also set AllowMajorVersionUpgrade to true.

modifyReplicationInstance_preferredMaintenanceWindow :: Lens' ModifyReplicationInstance (Maybe Text)

The weekly time range (in UTC) during which system maintenance can occur, which might result in an outage. Changing this parameter does not result in an outage, except in the following situation, and the change is asynchronously applied as soon as possible. If moving this window to the current time, there must be at least 30 minutes between the current time and end of the window to ensure pending changes are applied.

Default: Uses existing setting

Format: ddd:hh24:mi-ddd:hh24:mi

Valid Days: Mon | Tue | Wed | Thu | Fri | Sat | Sun

Constraints: Must be at least 30 minutes

modifyReplicationInstance_replicationInstanceIdentifier :: Lens' ModifyReplicationInstance (Maybe Text)

The replication instance identifier. This parameter is stored as a lowercase string.

modifyReplicationInstance_replicationInstanceClass :: Lens' ModifyReplicationInstance (Maybe Text)

The compute and memory capacity of the replication instance as defined for the specified replication instance class. For example to specify the instance class dms.c4.large, set this parameter to "dms.c4.large".

For more information on the settings and capacities for the available replication instance classes, see Selecting the right DMS replication instance for your migration.

modifyReplicationInstance_allocatedStorage :: Lens' ModifyReplicationInstance (Maybe Int)

The amount of storage (in gigabytes) to be allocated for the replication instance.

modifyReplicationInstance_applyImmediately :: Lens' ModifyReplicationInstance (Maybe Bool)

Indicates whether the changes should be applied immediately or during the next maintenance window.

modifyReplicationInstance_autoMinorVersionUpgrade :: Lens' ModifyReplicationInstance (Maybe Bool)

A value that indicates that minor version upgrades are applied automatically to the replication instance during the maintenance window. Changing this parameter doesn't result in an outage, except in the case described following. The change is asynchronously applied as soon as possible.

An outage does result if these factors apply:

This parameter is set to true during the maintenance window.

A newer minor version is available.

DMS has enabled automatic patching for the given engine version.

modifyReplicationInstance_replicationInstanceArn :: Lens' ModifyReplicationInstance Text

The Amazon Resource Name (ARN) of the replication instance.

Destructuring the Response

data ModifyReplicationInstanceResponse

See: newModifyReplicationInstanceResponse smart constructor.

Constructors

ModifyReplicationInstanceResponse'

Fields

replicationInstance :: Maybe ReplicationInstance

The modified replication instance.

httpStatus :: Int

The response's http status code.

newModifyReplicationInstanceResponse

Arguments

:: Int
$sel:httpStatus:ModifyReplicationInstanceResponse'

-> ModifyReplicationInstanceResponse

Create a value of ModifyReplicationInstanceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:replicationInstance:ModifyReplicationInstanceResponse', modifyReplicationInstanceResponse_replicationInstance - The modified replication instance.

$sel:httpStatus:ModifyReplicationInstanceResponse', modifyReplicationInstanceResponse_httpStatus - The response's http status code.

Response Lenses

modifyReplicationInstanceResponse_replicationInstance :: Lens' ModifyReplicationInstanceResponse (Maybe ReplicationInstance)

The modified replication instance.

modifyReplicationInstanceResponse_httpStatus :: Lens' ModifyReplicationInstanceResponse Int

The response's http status code.