Network.AWS.DMS.ModifyReplicationTask

Description

Modifies the specified replication task.

You can't modify the task endpoints. The task must be stopped before you can modify it.

For more information about DMS tasks, see Working with Migration Tasks in the Database Migration Service User Guide.

Creating a Request

data ModifyReplicationTask

See: newModifyReplicationTask smart constructor.

Constructors

ModifyReplicationTask'

Fields

migrationType :: Maybe MigrationTypeValue

The migration type. Valid values: full-load | cdc | full-load-and-cdc

taskData :: Maybe Text

Supplemental information that the task requires to migrate the data for certain source and target endpoints. For more information, see Specifying Supplemental Data for Task Settings in the Database Migration Service User Guide.

replicationTaskSettings :: Maybe Text

JSON file that contains settings for the task, such as task metadata settings.

tableMappings :: Maybe Text

When using the CLI or boto3, provide the path of the JSON file that contains the table mappings. Precede the path with file://. For example, --table-mappings file://mappingfile.json. When working with the DMS API, provide the JSON as the parameter value.

cdcStartTime :: Maybe POSIX

Indicates the start time for a change data capture (CDC) operation. Use either CdcStartTime or CdcStartPosition to specify when you want a CDC operation to start. Specifying both values results in an error.

Timestamp Example: --cdc-start-time “2018-03-08T12:12:12”

cdcStopPosition :: Maybe Text

Indicates when you want a change data capture (CDC) operation to stop. The value can be either server time or commit time.

Server time example: --cdc-stop-position “server_time:2018-02-09T12:12:12”

Commit time example: --cdc-stop-position “commit_time: 2018-02-09T12:12:12 “

cdcStartPosition :: Maybe Text

Indicates when you want a change data capture (CDC) operation to start. Use either CdcStartPosition or CdcStartTime to specify when you want a CDC operation to start. Specifying both values results in an error.

The value can be in date, checkpoint, or LSN/SCN format.

Date Example: --cdc-start-position “2018-03-08T12:12:12”

Checkpoint Example: --cdc-start-position "checkpoint:V1mysql-bin-changelog.157832:1975:-1:2002:677883278264080:mysql-bin-changelog.157832:187600#93"

LSN Example: --cdc-start-position “mysql-bin-changelog.000024:373”

When you use this task setting with a source PostgreSQL database, a logical replication slot should already be created and associated with the source endpoint. You can verify this by setting the slotName extra connection attribute to the name of this logical replication slot. For more information, see Extra Connection Attributes When Using PostgreSQL as a Source for DMS.

replicationTaskIdentifier :: Maybe Text

The replication task identifier.

Constraints:

Must contain 1-255 alphanumeric characters or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

replicationTaskArn :: Text

The Amazon Resource Name (ARN) of the replication task.

newModifyReplicationTask

Arguments

:: Text
$sel:replicationTaskArn:ModifyReplicationTask'

-> ModifyReplicationTask

Create a value of ModifyReplicationTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:migrationType:ModifyReplicationTask', modifyReplicationTask_migrationType - The migration type. Valid values: full-load | cdc | full-load-and-cdc

$sel:taskData:ModifyReplicationTask', modifyReplicationTask_taskData - Supplemental information that the task requires to migrate the data for certain source and target endpoints. For more information, see Specifying Supplemental Data for Task Settings in the Database Migration Service User Guide.

$sel:replicationTaskSettings:ModifyReplicationTask', modifyReplicationTask_replicationTaskSettings - JSON file that contains settings for the task, such as task metadata settings.

$sel:tableMappings:ModifyReplicationTask', modifyReplicationTask_tableMappings - When using the CLI or boto3, provide the path of the JSON file that contains the table mappings. Precede the path with file://. For example, --table-mappings file://mappingfile.json. When working with the DMS API, provide the JSON as the parameter value.

$sel:cdcStartTime:ModifyReplicationTask', modifyReplicationTask_cdcStartTime - Indicates the start time for a change data capture (CDC) operation. Use either CdcStartTime or CdcStartPosition to specify when you want a CDC operation to start. Specifying both values results in an error.

Timestamp Example: --cdc-start-time “2018-03-08T12:12:12”

$sel:cdcStopPosition:ModifyReplicationTask', modifyReplicationTask_cdcStopPosition - Indicates when you want a change data capture (CDC) operation to stop. The value can be either server time or commit time.

Server time example: --cdc-stop-position “server_time:2018-02-09T12:12:12”

Commit time example: --cdc-stop-position “commit_time: 2018-02-09T12:12:12 “

$sel:cdcStartPosition:ModifyReplicationTask', modifyReplicationTask_cdcStartPosition - Indicates when you want a change data capture (CDC) operation to start. Use either CdcStartPosition or CdcStartTime to specify when you want a CDC operation to start. Specifying both values results in an error.

The value can be in date, checkpoint, or LSN/SCN format.

Date Example: --cdc-start-position “2018-03-08T12:12:12”

Checkpoint Example: --cdc-start-position "checkpoint:V1mysql-bin-changelog.157832:1975:-1:2002:677883278264080:mysql-bin-changelog.157832:187600#93"

LSN Example: --cdc-start-position “mysql-bin-changelog.000024:373”

When you use this task setting with a source PostgreSQL database, a logical replication slot should already be created and associated with the source endpoint. You can verify this by setting the slotName extra connection attribute to the name of this logical replication slot. For more information, see Extra Connection Attributes When Using PostgreSQL as a Source for DMS.

$sel:replicationTaskIdentifier:ModifyReplicationTask', modifyReplicationTask_replicationTaskIdentifier - The replication task identifier.

Constraints:

Must contain 1-255 alphanumeric characters or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

$sel:replicationTaskArn:ModifyReplicationTask', modifyReplicationTask_replicationTaskArn - The Amazon Resource Name (ARN) of the replication task.

Request Lenses

modifyReplicationTask_migrationType :: Lens' ModifyReplicationTask (Maybe MigrationTypeValue)

The migration type. Valid values: full-load | cdc | full-load-and-cdc

modifyReplicationTask_taskData :: Lens' ModifyReplicationTask (Maybe Text)

Supplemental information that the task requires to migrate the data for certain source and target endpoints. For more information, see Specifying Supplemental Data for Task Settings in the Database Migration Service User Guide.

modifyReplicationTask_replicationTaskSettings :: Lens' ModifyReplicationTask (Maybe Text)

JSON file that contains settings for the task, such as task metadata settings.

modifyReplicationTask_tableMappings :: Lens' ModifyReplicationTask (Maybe Text)

When using the CLI or boto3, provide the path of the JSON file that contains the table mappings. Precede the path with file://. For example, --table-mappings file://mappingfile.json. When working with the DMS API, provide the JSON as the parameter value.

modifyReplicationTask_cdcStartTime :: Lens' ModifyReplicationTask (Maybe UTCTime)

Indicates the start time for a change data capture (CDC) operation. Use either CdcStartTime or CdcStartPosition to specify when you want a CDC operation to start. Specifying both values results in an error.

Timestamp Example: --cdc-start-time “2018-03-08T12:12:12”

modifyReplicationTask_cdcStopPosition :: Lens' ModifyReplicationTask (Maybe Text)

Indicates when you want a change data capture (CDC) operation to stop. The value can be either server time or commit time.

Server time example: --cdc-stop-position “server_time:2018-02-09T12:12:12”

Commit time example: --cdc-stop-position “commit_time: 2018-02-09T12:12:12 “

modifyReplicationTask_cdcStartPosition :: Lens' ModifyReplicationTask (Maybe Text)

Indicates when you want a change data capture (CDC) operation to start. Use either CdcStartPosition or CdcStartTime to specify when you want a CDC operation to start. Specifying both values results in an error.

The value can be in date, checkpoint, or LSN/SCN format.

Date Example: --cdc-start-position “2018-03-08T12:12:12”

Checkpoint Example: --cdc-start-position "checkpoint:V1mysql-bin-changelog.157832:1975:-1:2002:677883278264080:mysql-bin-changelog.157832:187600#93"

LSN Example: --cdc-start-position “mysql-bin-changelog.000024:373”

When you use this task setting with a source PostgreSQL database, a logical replication slot should already be created and associated with the source endpoint. You can verify this by setting the slotName extra connection attribute to the name of this logical replication slot. For more information, see Extra Connection Attributes When Using PostgreSQL as a Source for DMS.

modifyReplicationTask_replicationTaskIdentifier :: Lens' ModifyReplicationTask (Maybe Text)

The replication task identifier.

Constraints:

Must contain 1-255 alphanumeric characters or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

modifyReplicationTask_replicationTaskArn :: Lens' ModifyReplicationTask Text

The Amazon Resource Name (ARN) of the replication task.

Destructuring the Response

data ModifyReplicationTaskResponse

See: newModifyReplicationTaskResponse smart constructor.

Constructors

ModifyReplicationTaskResponse'

Fields

replicationTask :: Maybe ReplicationTask

The replication task that was modified.

httpStatus :: Int

The response's http status code.

newModifyReplicationTaskResponse

Arguments

:: Int
$sel:httpStatus:ModifyReplicationTaskResponse'

-> ModifyReplicationTaskResponse

Create a value of ModifyReplicationTaskResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:replicationTask:ModifyReplicationTaskResponse', modifyReplicationTaskResponse_replicationTask - The replication task that was modified.

$sel:httpStatus:ModifyReplicationTaskResponse', modifyReplicationTaskResponse_httpStatus - The response's http status code.

Response Lenses

modifyReplicationTaskResponse_replicationTask :: Lens' ModifyReplicationTaskResponse (Maybe ReplicationTask)

The replication task that was modified.

modifyReplicationTaskResponse_httpStatus :: Lens' ModifyReplicationTaskResponse Int

The response's http status code.