Network.AWS.DMS.StartReplicationTask

Description

Starts the replication task.

For more information about DMS tasks, see Working with Migration Tasks in the Database Migration Service User Guide.

Creating a Request

data StartReplicationTask

See: newStartReplicationTask smart constructor.

Constructors

StartReplicationTask'

Fields

cdcStartTime :: Maybe POSIX

Indicates the start time for a change data capture (CDC) operation. Use either CdcStartTime or CdcStartPosition to specify when you want a CDC operation to start. Specifying both values results in an error.

Timestamp Example: --cdc-start-time “2018-03-08T12:12:12”

cdcStopPosition :: Maybe Text

Indicates when you want a change data capture (CDC) operation to stop. The value can be either server time or commit time.

Server time example: --cdc-stop-position “server_time:2018-02-09T12:12:12”

Commit time example: --cdc-stop-position “commit_time: 2018-02-09T12:12:12 “

cdcStartPosition :: Maybe Text

Indicates when you want a change data capture (CDC) operation to start. Use either CdcStartPosition or CdcStartTime to specify when you want a CDC operation to start. Specifying both values results in an error.

The value can be in date, checkpoint, or LSN/SCN format.

Date Example: --cdc-start-position “2018-03-08T12:12:12”

Checkpoint Example: --cdc-start-position "checkpoint:V1mysql-bin-changelog.157832:1975:-1:2002:677883278264080:mysql-bin-changelog.157832:187600#93"

LSN Example: --cdc-start-position “mysql-bin-changelog.000024:373”

When you use this task setting with a source PostgreSQL database, a logical replication slot should already be created and associated with the source endpoint. You can verify this by setting the slotName extra connection attribute to the name of this logical replication slot. For more information, see Extra Connection Attributes When Using PostgreSQL as a Source for DMS.

replicationTaskArn :: Text

The Amazon Resource Name (ARN) of the replication task to be started.

startReplicationTaskType :: StartReplicationTaskTypeValue

A type of replication task.

newStartReplicationTask

Arguments

:: Text
$sel:replicationTaskArn:StartReplicationTask'

-> StartReplicationTaskTypeValue
$sel:startReplicationTaskType:StartReplicationTask'

-> StartReplicationTask

Create a value of StartReplicationTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cdcStartTime:StartReplicationTask', startReplicationTask_cdcStartTime - Indicates the start time for a change data capture (CDC) operation. Use either CdcStartTime or CdcStartPosition to specify when you want a CDC operation to start. Specifying both values results in an error.

Timestamp Example: --cdc-start-time “2018-03-08T12:12:12”

$sel:cdcStopPosition:StartReplicationTask', startReplicationTask_cdcStopPosition - Indicates when you want a change data capture (CDC) operation to stop. The value can be either server time or commit time.

Server time example: --cdc-stop-position “server_time:2018-02-09T12:12:12”

Commit time example: --cdc-stop-position “commit_time: 2018-02-09T12:12:12 “

$sel:cdcStartPosition:StartReplicationTask', startReplicationTask_cdcStartPosition - Indicates when you want a change data capture (CDC) operation to start. Use either CdcStartPosition or CdcStartTime to specify when you want a CDC operation to start. Specifying both values results in an error.

The value can be in date, checkpoint, or LSN/SCN format.

Date Example: --cdc-start-position “2018-03-08T12:12:12”

Checkpoint Example: --cdc-start-position "checkpoint:V1mysql-bin-changelog.157832:1975:-1:2002:677883278264080:mysql-bin-changelog.157832:187600#93"

LSN Example: --cdc-start-position “mysql-bin-changelog.000024:373”

When you use this task setting with a source PostgreSQL database, a logical replication slot should already be created and associated with the source endpoint. You can verify this by setting the slotName extra connection attribute to the name of this logical replication slot. For more information, see Extra Connection Attributes When Using PostgreSQL as a Source for DMS.

$sel:replicationTaskArn:StartReplicationTask', startReplicationTask_replicationTaskArn - The Amazon Resource Name (ARN) of the replication task to be started.

$sel:startReplicationTaskType:StartReplicationTask', startReplicationTask_startReplicationTaskType - A type of replication task.

Request Lenses

startReplicationTask_cdcStartTime :: Lens' StartReplicationTask (Maybe UTCTime)

Indicates the start time for a change data capture (CDC) operation. Use either CdcStartTime or CdcStartPosition to specify when you want a CDC operation to start. Specifying both values results in an error.

Timestamp Example: --cdc-start-time “2018-03-08T12:12:12”

startReplicationTask_cdcStopPosition :: Lens' StartReplicationTask (Maybe Text)

Indicates when you want a change data capture (CDC) operation to stop. The value can be either server time or commit time.

Server time example: --cdc-stop-position “server_time:2018-02-09T12:12:12”

Commit time example: --cdc-stop-position “commit_time: 2018-02-09T12:12:12 “

startReplicationTask_cdcStartPosition :: Lens' StartReplicationTask (Maybe Text)

Indicates when you want a change data capture (CDC) operation to start. Use either CdcStartPosition or CdcStartTime to specify when you want a CDC operation to start. Specifying both values results in an error.

The value can be in date, checkpoint, or LSN/SCN format.

Date Example: --cdc-start-position “2018-03-08T12:12:12”

Checkpoint Example: --cdc-start-position "checkpoint:V1mysql-bin-changelog.157832:1975:-1:2002:677883278264080:mysql-bin-changelog.157832:187600#93"

LSN Example: --cdc-start-position “mysql-bin-changelog.000024:373”

When you use this task setting with a source PostgreSQL database, a logical replication slot should already be created and associated with the source endpoint. You can verify this by setting the slotName extra connection attribute to the name of this logical replication slot. For more information, see Extra Connection Attributes When Using PostgreSQL as a Source for DMS.

startReplicationTask_replicationTaskArn :: Lens' StartReplicationTask Text

The Amazon Resource Name (ARN) of the replication task to be started.

startReplicationTask_startReplicationTaskType :: Lens' StartReplicationTask StartReplicationTaskTypeValue

A type of replication task.

Destructuring the Response

data StartReplicationTaskResponse

See: newStartReplicationTaskResponse smart constructor.

Constructors

StartReplicationTaskResponse'

Fields

replicationTask :: Maybe ReplicationTask

The replication task started.

httpStatus :: Int

The response's http status code.

newStartReplicationTaskResponse

Arguments

:: Int
$sel:httpStatus:StartReplicationTaskResponse'

-> StartReplicationTaskResponse

Create a value of StartReplicationTaskResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:replicationTask:StartReplicationTaskResponse', startReplicationTaskResponse_replicationTask - The replication task started.

$sel:httpStatus:StartReplicationTaskResponse', startReplicationTaskResponse_httpStatus - The response's http status code.

Response Lenses

startReplicationTaskResponse_replicationTask :: Lens' StartReplicationTaskResponse (Maybe ReplicationTask)

The replication task started.

startReplicationTaskResponse_httpStatus :: Lens' StartReplicationTaskResponse Int

The response's http status code.