Network.AWS.DMS.StartReplicationTaskAssessment

Description

Starts the replication task assessment for unsupported data types in the source database.

Creating a Request

data StartReplicationTaskAssessment

See: newStartReplicationTaskAssessment smart constructor.

Constructors

StartReplicationTaskAssessment'

Fields

replicationTaskArn :: Text

The Amazon Resource Name (ARN) of the replication task.

newStartReplicationTaskAssessment

Arguments

:: Text
$sel:replicationTaskArn:StartReplicationTaskAssessment'

-> StartReplicationTaskAssessment

Create a value of StartReplicationTaskAssessment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:replicationTaskArn:StartReplicationTaskAssessment', startReplicationTaskAssessment_replicationTaskArn - The Amazon Resource Name (ARN) of the replication task.

Request Lenses

startReplicationTaskAssessment_replicationTaskArn :: Lens' StartReplicationTaskAssessment Text

The Amazon Resource Name (ARN) of the replication task.

Destructuring the Response

data StartReplicationTaskAssessmentResponse

See: newStartReplicationTaskAssessmentResponse smart constructor.

Constructors

StartReplicationTaskAssessmentResponse'

Fields

replicationTask :: Maybe ReplicationTask

The assessed replication task.

httpStatus :: Int

The response's http status code.

newStartReplicationTaskAssessmentResponse

Arguments

:: Int
$sel:httpStatus:StartReplicationTaskAssessmentResponse'

-> StartReplicationTaskAssessmentResponse

Create a value of StartReplicationTaskAssessmentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:replicationTask:StartReplicationTaskAssessmentResponse', startReplicationTaskAssessmentResponse_replicationTask - The assessed replication task.

$sel:httpStatus:StartReplicationTaskAssessmentResponse', startReplicationTaskAssessmentResponse_httpStatus - The response's http status code.

Response Lenses

startReplicationTaskAssessmentResponse_replicationTask :: Lens' StartReplicationTaskAssessmentResponse (Maybe ReplicationTask)

The assessed replication task.

startReplicationTaskAssessmentResponse_httpStatus :: Lens' StartReplicationTaskAssessmentResponse Int

The response's http status code.