Network.AWS.DMS.StartReplicationTaskAssessmentRun

Description

Starts a new premigration assessment run for one or more individual assessments of a migration task.

The assessments that you can specify depend on the source and target database engine and the migration type defined for the given task. To run this operation, your migration task must already be created. After you run this operation, you can review the status of each individual assessment. You can also run the migration task manually after the assessment run and its individual assessments complete.

Creating a Request

data StartReplicationTaskAssessmentRun

See: newStartReplicationTaskAssessmentRun smart constructor.

Constructors

StartReplicationTaskAssessmentRun'

Fields

resultKmsKeyArn :: Maybe Text

ARN of a custom KMS encryption key that you specify when you set ResultEncryptionMode to "SSE_KMS".

resultEncryptionMode :: Maybe Text

Encryption mode that you can specify to encrypt the results of this assessment run. If you don't specify this request parameter, DMS stores the assessment run results without encryption. You can specify one of the options following:

"SSE_S3" – The server-side encryption provided as a default by Amazon S3.

"SSE_KMS" – Key Management Service (KMS) encryption. This encryption can use either a custom KMS encryption key that you specify or the default KMS encryption key that DMS provides.

exclude :: Maybe [Text]

Space-separated list of names for specific individual assessments that you want to exclude. These names come from the default list of individual assessments that DMS supports for the associated migration task. This task is specified by ReplicationTaskArn.

You can't set a value for Exclude if you also set a value for IncludeOnly in the API operation.

To identify the names of the default individual assessments that DMS supports for the associated migration task, run the DescribeApplicableIndividualAssessments operation using its own ReplicationTaskArn request parameter.

includeOnly :: Maybe [Text]

Space-separated list of names for specific individual assessments that you want to include. These names come from the default list of individual assessments that DMS supports for the associated migration task. This task is specified by ReplicationTaskArn.

You can't set a value for IncludeOnly if you also set a value for Exclude in the API operation.

To identify the names of the default individual assessments that DMS supports for the associated migration task, run the DescribeApplicableIndividualAssessments operation using its own ReplicationTaskArn request parameter.

resultLocationFolder :: Maybe Text

Folder within an Amazon S3 bucket where you want DMS to store the results of this assessment run.

replicationTaskArn :: Text

Amazon Resource Name (ARN) of the migration task associated with the premigration assessment run that you want to start.

serviceAccessRoleArn :: Text

ARN of the service role needed to start the assessment run. The role must allow the iam:PassRole action.

resultLocationBucket :: Text

Amazon S3 bucket where you want DMS to store the results of this assessment run.

assessmentRunName :: Text

Unique name to identify the assessment run.

newStartReplicationTaskAssessmentRun

Arguments

:: Text
$sel:replicationTaskArn:StartReplicationTaskAssessmentRun'

-> Text
$sel:serviceAccessRoleArn:StartReplicationTaskAssessmentRun'

-> Text
$sel:resultLocationBucket:StartReplicationTaskAssessmentRun'

-> Text
$sel:assessmentRunName:StartReplicationTaskAssessmentRun'

-> StartReplicationTaskAssessmentRun

Create a value of StartReplicationTaskAssessmentRun with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resultKmsKeyArn:StartReplicationTaskAssessmentRun', startReplicationTaskAssessmentRun_resultKmsKeyArn - ARN of a custom KMS encryption key that you specify when you set ResultEncryptionMode to "SSE_KMS".

$sel:resultEncryptionMode:StartReplicationTaskAssessmentRun', startReplicationTaskAssessmentRun_resultEncryptionMode - Encryption mode that you can specify to encrypt the results of this assessment run. If you don't specify this request parameter, DMS stores the assessment run results without encryption. You can specify one of the options following:

"SSE_S3" – The server-side encryption provided as a default by Amazon S3.

"SSE_KMS" – Key Management Service (KMS) encryption. This encryption can use either a custom KMS encryption key that you specify or the default KMS encryption key that DMS provides.

$sel:exclude:StartReplicationTaskAssessmentRun', startReplicationTaskAssessmentRun_exclude - Space-separated list of names for specific individual assessments that you want to exclude. These names come from the default list of individual assessments that DMS supports for the associated migration task. This task is specified by ReplicationTaskArn.

You can't set a value for Exclude if you also set a value for IncludeOnly in the API operation.

To identify the names of the default individual assessments that DMS supports for the associated migration task, run the DescribeApplicableIndividualAssessments operation using its own ReplicationTaskArn request parameter.

$sel:includeOnly:StartReplicationTaskAssessmentRun', startReplicationTaskAssessmentRun_includeOnly - Space-separated list of names for specific individual assessments that you want to include. These names come from the default list of individual assessments that DMS supports for the associated migration task. This task is specified by ReplicationTaskArn.

You can't set a value for IncludeOnly if you also set a value for Exclude in the API operation.

To identify the names of the default individual assessments that DMS supports for the associated migration task, run the DescribeApplicableIndividualAssessments operation using its own ReplicationTaskArn request parameter.

$sel:resultLocationFolder:StartReplicationTaskAssessmentRun', startReplicationTaskAssessmentRun_resultLocationFolder - Folder within an Amazon S3 bucket where you want DMS to store the results of this assessment run.

$sel:replicationTaskArn:StartReplicationTaskAssessmentRun', startReplicationTaskAssessmentRun_replicationTaskArn - Amazon Resource Name (ARN) of the migration task associated with the premigration assessment run that you want to start.

$sel:serviceAccessRoleArn:StartReplicationTaskAssessmentRun', startReplicationTaskAssessmentRun_serviceAccessRoleArn - ARN of the service role needed to start the assessment run. The role must allow the iam:PassRole action.

$sel:resultLocationBucket:StartReplicationTaskAssessmentRun', startReplicationTaskAssessmentRun_resultLocationBucket - Amazon S3 bucket where you want DMS to store the results of this assessment run.

$sel:assessmentRunName:StartReplicationTaskAssessmentRun', startReplicationTaskAssessmentRun_assessmentRunName - Unique name to identify the assessment run.

Request Lenses

startReplicationTaskAssessmentRun_resultKmsKeyArn :: Lens' StartReplicationTaskAssessmentRun (Maybe Text)

ARN of a custom KMS encryption key that you specify when you set ResultEncryptionMode to "SSE_KMS".

startReplicationTaskAssessmentRun_resultEncryptionMode :: Lens' StartReplicationTaskAssessmentRun (Maybe Text)

Encryption mode that you can specify to encrypt the results of this assessment run. If you don't specify this request parameter, DMS stores the assessment run results without encryption. You can specify one of the options following:

"SSE_S3" – The server-side encryption provided as a default by Amazon S3.

"SSE_KMS" – Key Management Service (KMS) encryption. This encryption can use either a custom KMS encryption key that you specify or the default KMS encryption key that DMS provides.

startReplicationTaskAssessmentRun_exclude :: Lens' StartReplicationTaskAssessmentRun (Maybe [Text])

Space-separated list of names for specific individual assessments that you want to exclude. These names come from the default list of individual assessments that DMS supports for the associated migration task. This task is specified by ReplicationTaskArn.

You can't set a value for Exclude if you also set a value for IncludeOnly in the API operation.

To identify the names of the default individual assessments that DMS supports for the associated migration task, run the DescribeApplicableIndividualAssessments operation using its own ReplicationTaskArn request parameter.

startReplicationTaskAssessmentRun_includeOnly :: Lens' StartReplicationTaskAssessmentRun (Maybe [Text])

Space-separated list of names for specific individual assessments that you want to include. These names come from the default list of individual assessments that DMS supports for the associated migration task. This task is specified by ReplicationTaskArn.

You can't set a value for IncludeOnly if you also set a value for Exclude in the API operation.

To identify the names of the default individual assessments that DMS supports for the associated migration task, run the DescribeApplicableIndividualAssessments operation using its own ReplicationTaskArn request parameter.

startReplicationTaskAssessmentRun_resultLocationFolder :: Lens' StartReplicationTaskAssessmentRun (Maybe Text)

Folder within an Amazon S3 bucket where you want DMS to store the results of this assessment run.

startReplicationTaskAssessmentRun_replicationTaskArn :: Lens' StartReplicationTaskAssessmentRun Text

Amazon Resource Name (ARN) of the migration task associated with the premigration assessment run that you want to start.

startReplicationTaskAssessmentRun_serviceAccessRoleArn :: Lens' StartReplicationTaskAssessmentRun Text

ARN of the service role needed to start the assessment run. The role must allow the iam:PassRole action.

startReplicationTaskAssessmentRun_resultLocationBucket :: Lens' StartReplicationTaskAssessmentRun Text

Amazon S3 bucket where you want DMS to store the results of this assessment run.

startReplicationTaskAssessmentRun_assessmentRunName :: Lens' StartReplicationTaskAssessmentRun Text

Unique name to identify the assessment run.

Destructuring the Response

data StartReplicationTaskAssessmentRunResponse

See: newStartReplicationTaskAssessmentRunResponse smart constructor.

Constructors

StartReplicationTaskAssessmentRunResponse'

Fields

replicationTaskAssessmentRun :: Maybe ReplicationTaskAssessmentRun

The premigration assessment run that was started.

httpStatus :: Int

The response's http status code.

newStartReplicationTaskAssessmentRunResponse

Arguments

:: Int
$sel:httpStatus:StartReplicationTaskAssessmentRunResponse'

-> StartReplicationTaskAssessmentRunResponse

Create a value of StartReplicationTaskAssessmentRunResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:replicationTaskAssessmentRun:StartReplicationTaskAssessmentRunResponse', startReplicationTaskAssessmentRunResponse_replicationTaskAssessmentRun - The premigration assessment run that was started.

$sel:httpStatus:StartReplicationTaskAssessmentRunResponse', startReplicationTaskAssessmentRunResponse_httpStatus - The response's http status code.

Response Lenses

startReplicationTaskAssessmentRunResponse_replicationTaskAssessmentRun :: Lens' StartReplicationTaskAssessmentRunResponse (Maybe ReplicationTaskAssessmentRun)

The premigration assessment run that was started.

startReplicationTaskAssessmentRunResponse_httpStatus :: Lens' StartReplicationTaskAssessmentRunResponse Int

The response's http status code.