Network.AWS.DMS.Types.Connection

Description

Documentation

data Connection

Status of the connection between an endpoint and a replication instance, including Amazon Resource Names (ARNs) and the last error message issued.

See: newConnection smart constructor.

Constructors

Connection'

Fields

status :: Maybe Text

The connection status. This parameter can return one of the following values:

"successful"

"testing"

"failed"

"deleting"

replicationInstanceArn :: Maybe Text

The ARN of the replication instance.

endpointIdentifier :: Maybe Text

The identifier of the endpoint. Identifiers must begin with a letter and must contain only ASCII letters, digits, and hyphens. They can't end with a hyphen or contain two consecutive hyphens.

replicationInstanceIdentifier :: Maybe Text

The replication instance identifier. This parameter is stored as a lowercase string.

endpointArn :: Maybe Text

The ARN string that uniquely identifies the endpoint.

lastFailureMessage :: Maybe Text

The error message when the connection last failed.

newConnection :: Connection

Create a value of Connection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Connection', connection_status - The connection status. This parameter can return one of the following values:

"successful"

"testing"

"failed"

"deleting"

$sel:replicationInstanceArn:Connection', connection_replicationInstanceArn - The ARN of the replication instance.

$sel:endpointIdentifier:Connection', connection_endpointIdentifier - The identifier of the endpoint. Identifiers must begin with a letter and must contain only ASCII letters, digits, and hyphens. They can't end with a hyphen or contain two consecutive hyphens.

$sel:replicationInstanceIdentifier:Connection', connection_replicationInstanceIdentifier - The replication instance identifier. This parameter is stored as a lowercase string.

$sel:endpointArn:Connection', connection_endpointArn - The ARN string that uniquely identifies the endpoint.

$sel:lastFailureMessage:Connection', connection_lastFailureMessage - The error message when the connection last failed.

connection_status :: Lens' Connection (Maybe Text)

The connection status. This parameter can return one of the following values:

"successful"

"testing"

"failed"

"deleting"

connection_replicationInstanceArn :: Lens' Connection (Maybe Text)

The ARN of the replication instance.

connection_endpointIdentifier :: Lens' Connection (Maybe Text)

The identifier of the endpoint. Identifiers must begin with a letter and must contain only ASCII letters, digits, and hyphens. They can't end with a hyphen or contain two consecutive hyphens.

connection_replicationInstanceIdentifier :: Lens' Connection (Maybe Text)

The replication instance identifier. This parameter is stored as a lowercase string.

connection_endpointArn :: Lens' Connection (Maybe Text)

The ARN string that uniquely identifies the endpoint.

connection_lastFailureMessage :: Lens' Connection (Maybe Text)

The error message when the connection last failed.