Network.AWS.DMS.Types.ElasticsearchSettings

Description

Documentation

data ElasticsearchSettings

Provides information that defines an Elasticsearch endpoint.

See: newElasticsearchSettings smart constructor.

Constructors

ElasticsearchSettings'

Fields

fullLoadErrorPercentage :: Maybe Int

The maximum percentage of records that can fail to be written before a full load operation stops.

To avoid early failure, this counter is only effective after 1000 records are transferred. Elasticsearch also has the concept of error monitoring during the last 10 minutes of an Observation Window. If transfer of all records fail in the last 10 minutes, the full load operation stops.

errorRetryDuration :: Maybe Int

The maximum number of seconds for which DMS retries failed API requests to the Elasticsearch cluster.

serviceAccessRoleArn :: Text

The Amazon Resource Name (ARN) used by the service to access the IAM role. The role must allow the iam:PassRole action.

endpointUri :: Text

The endpoint for the Elasticsearch cluster. DMS uses HTTPS if a transport protocol (http/https) is not specified.

newElasticsearchSettings

Arguments

:: Text
$sel:serviceAccessRoleArn:ElasticsearchSettings'

-> Text
$sel:endpointUri:ElasticsearchSettings'

-> ElasticsearchSettings

Create a value of ElasticsearchSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fullLoadErrorPercentage:ElasticsearchSettings', elasticsearchSettings_fullLoadErrorPercentage - The maximum percentage of records that can fail to be written before a full load operation stops.

To avoid early failure, this counter is only effective after 1000 records are transferred. Elasticsearch also has the concept of error monitoring during the last 10 minutes of an Observation Window. If transfer of all records fail in the last 10 minutes, the full load operation stops.

$sel:errorRetryDuration:ElasticsearchSettings', elasticsearchSettings_errorRetryDuration - The maximum number of seconds for which DMS retries failed API requests to the Elasticsearch cluster.

$sel:serviceAccessRoleArn:ElasticsearchSettings', elasticsearchSettings_serviceAccessRoleArn - The Amazon Resource Name (ARN) used by the service to access the IAM role. The role must allow the iam:PassRole action.

$sel:endpointUri:ElasticsearchSettings', elasticsearchSettings_endpointUri - The endpoint for the Elasticsearch cluster. DMS uses HTTPS if a transport protocol (http/https) is not specified.

elasticsearchSettings_fullLoadErrorPercentage :: Lens' ElasticsearchSettings (Maybe Int)

The maximum percentage of records that can fail to be written before a full load operation stops.

To avoid early failure, this counter is only effective after 1000 records are transferred. Elasticsearch also has the concept of error monitoring during the last 10 minutes of an Observation Window. If transfer of all records fail in the last 10 minutes, the full load operation stops.

elasticsearchSettings_errorRetryDuration :: Lens' ElasticsearchSettings (Maybe Int)

The maximum number of seconds for which DMS retries failed API requests to the Elasticsearch cluster.

elasticsearchSettings_serviceAccessRoleArn :: Lens' ElasticsearchSettings Text

The Amazon Resource Name (ARN) used by the service to access the IAM role. The role must allow the iam:PassRole action.

elasticsearchSettings_endpointUri :: Lens' ElasticsearchSettings Text

The endpoint for the Elasticsearch cluster. DMS uses HTTPS if a transport protocol (http/https) is not specified.