Network.AWS.DMS.Types.Endpoint

Description

Documentation

data Endpoint

Describes an endpoint of a database instance in response to operations such as the following:

CreateEndpoint

DescribeEndpoint

ModifyEndpoint

See: newEndpoint smart constructor.

Constructors

Endpoint'

Fields

status :: Maybe Text

The status of the endpoint.

dmsTransferSettings :: Maybe DmsTransferSettings

The settings in JSON format for the DMS transfer type of source endpoint.

Possible settings include the following:

ServiceAccessRoleArn - - The Amazon Resource Name (ARN) used by the service access IAM role. The role must allow the iam:PassRole action.

BucketName - The name of the S3 bucket to use.

Shorthand syntax for these settings is as follows: ServiceAccessRoleArn=string,BucketName=string,

JSON syntax for these settings is as follows: { "ServiceAccessRoleArn": "string", "BucketName": "string"}

mySQLSettings :: Maybe MySQLSettings

The settings for the MySQL source and target endpoint. For more information, see the MySQLSettings structure.

serverName :: Maybe Text

The name of the server at the endpoint.

microsoftSQLServerSettings :: Maybe MicrosoftSQLServerSettings

The settings for the Microsoft SQL Server source and target endpoint. For more information, see the MicrosoftSQLServerSettings structure.

certificateArn :: Maybe Text

The Amazon Resource Name (ARN) used for SSL connection to the endpoint.

serviceAccessRoleArn :: Maybe Text

The Amazon Resource Name (ARN) used by the service to access the IAM role. The role must allow the iam:PassRole action.

docDbSettings :: Maybe DocDbSettings

engineDisplayName :: Maybe Text

The expanded name for the engine name. For example, if the EngineName parameter is "aurora," this value would be "Amazon Aurora MySQL."

postgreSQLSettings :: Maybe PostgreSQLSettings

The settings for the PostgreSQL source and target endpoint. For more information, see the PostgreSQLSettings structure.

extraConnectionAttributes :: Maybe Text

Additional connection attributes used to connect to the endpoint.

kafkaSettings :: Maybe KafkaSettings

The settings for the Apache Kafka target endpoint. For more information, see the KafkaSettings structure.

oracleSettings :: Maybe OracleSettings

The settings for the Oracle source and target endpoint. For more information, see the OracleSettings structure.

endpointType :: Maybe ReplicationEndpointTypeValue

The type of endpoint. Valid values are source and target.

redshiftSettings :: Maybe RedshiftSettings

Settings for the Amazon Redshift endpoint.

elasticsearchSettings :: Maybe ElasticsearchSettings

The settings for the Elasticsearch source endpoint. For more information, see the ElasticsearchSettings structure.

username :: Maybe Text

The user name used to connect to the endpoint.

externalTableDefinition :: Maybe Text

The external table definition.

engineName :: Maybe Text

The database engine name. Valid values, depending on the EndpointType, include "mysql", "oracle", "postgres", "mariadb", "aurora", "aurora-postgresql", "redshift", "s3", "db2", "azuredb", "sybase", "dynamodb", "mongodb", "kinesis", "kafka", "elasticsearch", "documentdb", "sqlserver", and "neptune".

redisSettings :: Maybe RedisSettings

The settings for the Redis target endpoint. For more information, see the RedisSettings structure.

neptuneSettings :: Maybe NeptuneSettings

The settings for the Amazon Neptune target endpoint. For more information, see the NeptuneSettings structure.

iBMDb2Settings :: Maybe IBMDb2Settings

The settings for the IBM Db2 LUW source endpoint. For more information, see the IBMDb2Settings structure.

kmsKeyId :: Maybe Text

An KMS key identifier that is used to encrypt the connection parameters for the endpoint.

If you don't specify a value for the KmsKeyId parameter, then DMS uses your default encryption key.

KMS creates the default encryption key for your Amazon Web Services account. Your Amazon Web Services account has a different default encryption key for each Amazon Web Services Region.

mongoDbSettings :: Maybe MongoDbSettings

The settings for the MongoDB source endpoint. For more information, see the MongoDbSettings structure.

sslMode :: Maybe DmsSslModeValue

The SSL mode used to connect to the endpoint. The default value is none.

sybaseSettings :: Maybe SybaseSettings

The settings for the SAP ASE source and target endpoint. For more information, see the SybaseSettings structure.

databaseName :: Maybe Text

The name of the database at the endpoint.

s3Settings :: Maybe S3Settings

The settings for the S3 target endpoint. For more information, see the S3Settings structure.

kinesisSettings :: Maybe KinesisSettings

The settings for the Amazon Kinesis target endpoint. For more information, see the KinesisSettings structure.

endpointIdentifier :: Maybe Text

The database endpoint identifier. Identifiers must begin with a letter and must contain only ASCII letters, digits, and hyphens. They can't end with a hyphen or contain two consecutive hyphens.

externalId :: Maybe Text

Value returned by a call to CreateEndpoint that can be used for cross-account validation. Use it on a subsequent call to CreateEndpoint to create the endpoint with a cross-account.

dynamoDbSettings :: Maybe DynamoDbSettings

The settings for the DynamoDB target endpoint. For more information, see the DynamoDBSettings structure.

endpointArn :: Maybe Text

The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.

port :: Maybe Int

The port value used to access the endpoint.

newEndpoint :: Endpoint

Create a value of Endpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Endpoint', endpoint_status - The status of the endpoint.

$sel:dmsTransferSettings:Endpoint', endpoint_dmsTransferSettings - The settings in JSON format for the DMS transfer type of source endpoint.

Possible settings include the following:

ServiceAccessRoleArn - - The Amazon Resource Name (ARN) used by the service access IAM role. The role must allow the iam:PassRole action.

BucketName - The name of the S3 bucket to use.

Shorthand syntax for these settings is as follows: ServiceAccessRoleArn=string,BucketName=string,

JSON syntax for these settings is as follows: { "ServiceAccessRoleArn": "string", "BucketName": "string"}

$sel:mySQLSettings:Endpoint', endpoint_mySQLSettings - The settings for the MySQL source and target endpoint. For more information, see the MySQLSettings structure.

$sel:serverName:Endpoint', endpoint_serverName - The name of the server at the endpoint.

$sel:microsoftSQLServerSettings:Endpoint', endpoint_microsoftSQLServerSettings - The settings for the Microsoft SQL Server source and target endpoint. For more information, see the MicrosoftSQLServerSettings structure.

$sel:certificateArn:Endpoint', endpoint_certificateArn - The Amazon Resource Name (ARN) used for SSL connection to the endpoint.

$sel:serviceAccessRoleArn:Endpoint', endpoint_serviceAccessRoleArn - The Amazon Resource Name (ARN) used by the service to access the IAM role. The role must allow the iam:PassRole action.

$sel:docDbSettings:Endpoint', endpoint_docDbSettings - Undocumented member.

$sel:engineDisplayName:Endpoint', endpoint_engineDisplayName - The expanded name for the engine name. For example, if the EngineName parameter is "aurora," this value would be "Amazon Aurora MySQL."

$sel:postgreSQLSettings:Endpoint', endpoint_postgreSQLSettings - The settings for the PostgreSQL source and target endpoint. For more information, see the PostgreSQLSettings structure.

$sel:extraConnectionAttributes:Endpoint', endpoint_extraConnectionAttributes - Additional connection attributes used to connect to the endpoint.

$sel:kafkaSettings:Endpoint', endpoint_kafkaSettings - The settings for the Apache Kafka target endpoint. For more information, see the KafkaSettings structure.

$sel:oracleSettings:Endpoint', endpoint_oracleSettings - The settings for the Oracle source and target endpoint. For more information, see the OracleSettings structure.

$sel:endpointType:Endpoint', endpoint_endpointType - The type of endpoint. Valid values are source and target.

$sel:redshiftSettings:Endpoint', endpoint_redshiftSettings - Settings for the Amazon Redshift endpoint.

$sel:elasticsearchSettings:Endpoint', endpoint_elasticsearchSettings - The settings for the Elasticsearch source endpoint. For more information, see the ElasticsearchSettings structure.

$sel:username:Endpoint', endpoint_username - The user name used to connect to the endpoint.

$sel:externalTableDefinition:Endpoint', endpoint_externalTableDefinition - The external table definition.

$sel:engineName:Endpoint', endpoint_engineName - The database engine name. Valid values, depending on the EndpointType, include "mysql", "oracle", "postgres", "mariadb", "aurora", "aurora-postgresql", "redshift", "s3", "db2", "azuredb", "sybase", "dynamodb", "mongodb", "kinesis", "kafka", "elasticsearch", "documentdb", "sqlserver", and "neptune".

$sel:redisSettings:Endpoint', endpoint_redisSettings - The settings for the Redis target endpoint. For more information, see the RedisSettings structure.

$sel:neptuneSettings:Endpoint', endpoint_neptuneSettings - The settings for the Amazon Neptune target endpoint. For more information, see the NeptuneSettings structure.

$sel:iBMDb2Settings:Endpoint', endpoint_iBMDb2Settings - The settings for the IBM Db2 LUW source endpoint. For more information, see the IBMDb2Settings structure.

$sel:kmsKeyId:Endpoint', endpoint_kmsKeyId - An KMS key identifier that is used to encrypt the connection parameters for the endpoint.

If you don't specify a value for the KmsKeyId parameter, then DMS uses your default encryption key.

KMS creates the default encryption key for your Amazon Web Services account. Your Amazon Web Services account has a different default encryption key for each Amazon Web Services Region.

$sel:mongoDbSettings:Endpoint', endpoint_mongoDbSettings - The settings for the MongoDB source endpoint. For more information, see the MongoDbSettings structure.

$sel:sslMode:Endpoint', endpoint_sslMode - The SSL mode used to connect to the endpoint. The default value is none.

$sel:sybaseSettings:Endpoint', endpoint_sybaseSettings - The settings for the SAP ASE source and target endpoint. For more information, see the SybaseSettings structure.

$sel:databaseName:Endpoint', endpoint_databaseName - The name of the database at the endpoint.

$sel:s3Settings:Endpoint', endpoint_s3Settings - The settings for the S3 target endpoint. For more information, see the S3Settings structure.

$sel:kinesisSettings:Endpoint', endpoint_kinesisSettings - The settings for the Amazon Kinesis target endpoint. For more information, see the KinesisSettings structure.

$sel:endpointIdentifier:Endpoint', endpoint_endpointIdentifier - The database endpoint identifier. Identifiers must begin with a letter and must contain only ASCII letters, digits, and hyphens. They can't end with a hyphen or contain two consecutive hyphens.

$sel:externalId:Endpoint', endpoint_externalId - Value returned by a call to CreateEndpoint that can be used for cross-account validation. Use it on a subsequent call to CreateEndpoint to create the endpoint with a cross-account.

$sel:dynamoDbSettings:Endpoint', endpoint_dynamoDbSettings - The settings for the DynamoDB target endpoint. For more information, see the DynamoDBSettings structure.

$sel:endpointArn:Endpoint', endpoint_endpointArn - The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.

$sel:port:Endpoint', endpoint_port - The port value used to access the endpoint.

endpoint_status :: Lens' Endpoint (Maybe Text)

The status of the endpoint.

endpoint_dmsTransferSettings :: Lens' Endpoint (Maybe DmsTransferSettings)

The settings in JSON format for the DMS transfer type of source endpoint.

Possible settings include the following:

ServiceAccessRoleArn - - The Amazon Resource Name (ARN) used by the service access IAM role. The role must allow the iam:PassRole action.

BucketName - The name of the S3 bucket to use.

Shorthand syntax for these settings is as follows: ServiceAccessRoleArn=string,BucketName=string,

JSON syntax for these settings is as follows: { "ServiceAccessRoleArn": "string", "BucketName": "string"}

endpoint_mySQLSettings :: Lens' Endpoint (Maybe MySQLSettings)

The settings for the MySQL source and target endpoint. For more information, see the MySQLSettings structure.

endpoint_serverName :: Lens' Endpoint (Maybe Text)

The name of the server at the endpoint.

endpoint_microsoftSQLServerSettings :: Lens' Endpoint (Maybe MicrosoftSQLServerSettings)

The settings for the Microsoft SQL Server source and target endpoint. For more information, see the MicrosoftSQLServerSettings structure.

endpoint_certificateArn :: Lens' Endpoint (Maybe Text)

The Amazon Resource Name (ARN) used for SSL connection to the endpoint.

endpoint_serviceAccessRoleArn :: Lens' Endpoint (Maybe Text)

The Amazon Resource Name (ARN) used by the service to access the IAM role. The role must allow the iam:PassRole action.

endpoint_docDbSettings :: Lens' Endpoint (Maybe DocDbSettings)

Undocumented member.

endpoint_engineDisplayName :: Lens' Endpoint (Maybe Text)

The expanded name for the engine name. For example, if the EngineName parameter is "aurora," this value would be "Amazon Aurora MySQL."

endpoint_postgreSQLSettings :: Lens' Endpoint (Maybe PostgreSQLSettings)

The settings for the PostgreSQL source and target endpoint. For more information, see the PostgreSQLSettings structure.

endpoint_extraConnectionAttributes :: Lens' Endpoint (Maybe Text)

Additional connection attributes used to connect to the endpoint.

endpoint_kafkaSettings :: Lens' Endpoint (Maybe KafkaSettings)

The settings for the Apache Kafka target endpoint. For more information, see the KafkaSettings structure.

endpoint_oracleSettings :: Lens' Endpoint (Maybe OracleSettings)

The settings for the Oracle source and target endpoint. For more information, see the OracleSettings structure.

endpoint_endpointType :: Lens' Endpoint (Maybe ReplicationEndpointTypeValue)

The type of endpoint. Valid values are source and target.

endpoint_redshiftSettings :: Lens' Endpoint (Maybe RedshiftSettings)

Settings for the Amazon Redshift endpoint.

endpoint_elasticsearchSettings :: Lens' Endpoint (Maybe ElasticsearchSettings)

The settings for the Elasticsearch source endpoint. For more information, see the ElasticsearchSettings structure.

endpoint_username :: Lens' Endpoint (Maybe Text)

The user name used to connect to the endpoint.

endpoint_externalTableDefinition :: Lens' Endpoint (Maybe Text)

The external table definition.

endpoint_engineName :: Lens' Endpoint (Maybe Text)

The database engine name. Valid values, depending on the EndpointType, include "mysql", "oracle", "postgres", "mariadb", "aurora", "aurora-postgresql", "redshift", "s3", "db2", "azuredb", "sybase", "dynamodb", "mongodb", "kinesis", "kafka", "elasticsearch", "documentdb", "sqlserver", and "neptune".

endpoint_redisSettings :: Lens' Endpoint (Maybe RedisSettings)

The settings for the Redis target endpoint. For more information, see the RedisSettings structure.

endpoint_neptuneSettings :: Lens' Endpoint (Maybe NeptuneSettings)

The settings for the Amazon Neptune target endpoint. For more information, see the NeptuneSettings structure.

endpoint_iBMDb2Settings :: Lens' Endpoint (Maybe IBMDb2Settings)

The settings for the IBM Db2 LUW source endpoint. For more information, see the IBMDb2Settings structure.

endpoint_kmsKeyId :: Lens' Endpoint (Maybe Text)

An KMS key identifier that is used to encrypt the connection parameters for the endpoint.

If you don't specify a value for the KmsKeyId parameter, then DMS uses your default encryption key.

KMS creates the default encryption key for your Amazon Web Services account. Your Amazon Web Services account has a different default encryption key for each Amazon Web Services Region.

endpoint_mongoDbSettings :: Lens' Endpoint (Maybe MongoDbSettings)

The settings for the MongoDB source endpoint. For more information, see the MongoDbSettings structure.

endpoint_sslMode :: Lens' Endpoint (Maybe DmsSslModeValue)

The SSL mode used to connect to the endpoint. The default value is none.

endpoint_sybaseSettings :: Lens' Endpoint (Maybe SybaseSettings)

The settings for the SAP ASE source and target endpoint. For more information, see the SybaseSettings structure.

endpoint_databaseName :: Lens' Endpoint (Maybe Text)

The name of the database at the endpoint.

endpoint_s3Settings :: Lens' Endpoint (Maybe S3Settings)

The settings for the S3 target endpoint. For more information, see the S3Settings structure.

endpoint_kinesisSettings :: Lens' Endpoint (Maybe KinesisSettings)

The settings for the Amazon Kinesis target endpoint. For more information, see the KinesisSettings structure.

endpoint_endpointIdentifier :: Lens' Endpoint (Maybe Text)

The database endpoint identifier. Identifiers must begin with a letter and must contain only ASCII letters, digits, and hyphens. They can't end with a hyphen or contain two consecutive hyphens.

endpoint_externalId :: Lens' Endpoint (Maybe Text)

Value returned by a call to CreateEndpoint that can be used for cross-account validation. Use it on a subsequent call to CreateEndpoint to create the endpoint with a cross-account.

endpoint_dynamoDbSettings :: Lens' Endpoint (Maybe DynamoDbSettings)

The settings for the DynamoDB target endpoint. For more information, see the DynamoDBSettings structure.

endpoint_endpointArn :: Lens' Endpoint (Maybe Text)

The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.

endpoint_port :: Lens' Endpoint (Maybe Int)

The port value used to access the endpoint.