Network.AWS.DMS.Types.Event

Description

Documentation

data Event

Describes an identifiable significant activity that affects a replication instance or task. This object can provide the message, the available event categories, the date and source of the event, and the DMS resource type.

See: newEvent smart constructor.

Constructors

Event'

Fields

message :: Maybe Text

The event message.

eventCategories :: Maybe [Text]

The event categories available for the specified source type.

date :: Maybe POSIX

The date of the event.

sourceIdentifier :: Maybe Text

The identifier of an event source.

sourceType :: Maybe SourceType

The type of DMS resource that generates events.

Valid values: replication-instance | endpoint | replication-task

newEvent :: Event

Create a value of Event with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:message:Event', event_message - The event message.

$sel:eventCategories:Event', event_eventCategories - The event categories available for the specified source type.

$sel:date:Event', event_date - The date of the event.

$sel:sourceIdentifier:Event', event_sourceIdentifier - The identifier of an event source.

$sel:sourceType:Event', event_sourceType - The type of DMS resource that generates events.

Valid values: replication-instance | endpoint | replication-task

event_message :: Lens' Event (Maybe Text)

The event message.

event_eventCategories :: Lens' Event (Maybe [Text])

The event categories available for the specified source type.

event_date :: Lens' Event (Maybe UTCTime)

The date of the event.

event_sourceIdentifier :: Lens' Event (Maybe Text)

The identifier of an event source.

event_sourceType :: Lens' Event (Maybe SourceType)

The type of DMS resource that generates events.

Valid values: replication-instance | endpoint | replication-task