Network.AWS.DMS.Types.IBMDb2Settings

Description

Documentation

data IBMDb2Settings

Provides information that defines an IBM Db2 LUW endpoint.

See: newIBMDb2Settings smart constructor.

Constructors

IBMDb2Settings'

Fields

currentLsn :: Maybe Text

For ongoing replication (CDC), use CurrentLSN to specify a log sequence number (LSN) where you want the replication to start.

secretsManagerSecretId :: Maybe Text

The full ARN, partial ARN, or friendly name of the SecretsManagerSecret that contains the Db2 LUW endpoint connection details.

serverName :: Maybe Text

Fully qualified domain name of the endpoint.

password :: Maybe (Sensitive Text)

Endpoint connection password.

port :: Maybe Int

Endpoint TCP port. The default value is 50000.

username :: Maybe Text

Endpoint connection user name.

secretsManagerAccessRoleArn :: Maybe Text

The full Amazon Resource Name (ARN) of the IAM role that specifies DMS as the trusted entity and grants the required permissions to access the value in SecretsManagerSecret. The role must allow the iam:PassRole action. SecretsManagerSecret has the value of the Amazon Web Services Secrets Manager secret that allows access to the Db2 LUW endpoint.

You can specify one of two sets of values for these permissions. You can specify the values for this setting and SecretsManagerSecretId. Or you can specify clear-text values for UserName, Password, ServerName, and Port. You can't specify both. For more information on creating this SecretsManagerSecret and the SecretsManagerAccessRoleArn and SecretsManagerSecretId required to access it, see Using secrets to access Database Migration Service resources in the Database Migration Service User Guide.

maxKBytesPerRead :: Maybe Int

Maximum number of bytes per read, as a NUMBER value. The default is 64 KB.

setDataCaptureChanges :: Maybe Bool

Enables ongoing replication (CDC) as a BOOLEAN value. The default is true.

databaseName :: Maybe Text

Database name for the endpoint.

newIBMDb2Settings :: IBMDb2Settings

Create a value of IBMDb2Settings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:currentLsn:IBMDb2Settings', iBMDb2Settings_currentLsn - For ongoing replication (CDC), use CurrentLSN to specify a log sequence number (LSN) where you want the replication to start.

$sel:secretsManagerSecretId:IBMDb2Settings', iBMDb2Settings_secretsManagerSecretId - The full ARN, partial ARN, or friendly name of the SecretsManagerSecret that contains the Db2 LUW endpoint connection details.

$sel:serverName:IBMDb2Settings', iBMDb2Settings_serverName - Fully qualified domain name of the endpoint.

$sel:password:IBMDb2Settings', iBMDb2Settings_password - Endpoint connection password.

$sel:port:IBMDb2Settings', iBMDb2Settings_port - Endpoint TCP port. The default value is 50000.

$sel:username:IBMDb2Settings', iBMDb2Settings_username - Endpoint connection user name.

$sel:secretsManagerAccessRoleArn:IBMDb2Settings', iBMDb2Settings_secretsManagerAccessRoleArn - The full Amazon Resource Name (ARN) of the IAM role that specifies DMS as the trusted entity and grants the required permissions to access the value in SecretsManagerSecret. The role must allow the iam:PassRole action. SecretsManagerSecret has the value of the Amazon Web Services Secrets Manager secret that allows access to the Db2 LUW endpoint.

You can specify one of two sets of values for these permissions. You can specify the values for this setting and SecretsManagerSecretId. Or you can specify clear-text values for UserName, Password, ServerName, and Port. You can't specify both. For more information on creating this SecretsManagerSecret and the SecretsManagerAccessRoleArn and SecretsManagerSecretId required to access it, see Using secrets to access Database Migration Service resources in the Database Migration Service User Guide.

$sel:maxKBytesPerRead:IBMDb2Settings', iBMDb2Settings_maxKBytesPerRead - Maximum number of bytes per read, as a NUMBER value. The default is 64 KB.

$sel:setDataCaptureChanges:IBMDb2Settings', iBMDb2Settings_setDataCaptureChanges - Enables ongoing replication (CDC) as a BOOLEAN value. The default is true.

$sel:databaseName:IBMDb2Settings', iBMDb2Settings_databaseName - Database name for the endpoint.

iBMDb2Settings_currentLsn :: Lens' IBMDb2Settings (Maybe Text)

For ongoing replication (CDC), use CurrentLSN to specify a log sequence number (LSN) where you want the replication to start.

iBMDb2Settings_secretsManagerSecretId :: Lens' IBMDb2Settings (Maybe Text)

The full ARN, partial ARN, or friendly name of the SecretsManagerSecret that contains the Db2 LUW endpoint connection details.

iBMDb2Settings_serverName :: Lens' IBMDb2Settings (Maybe Text)

Fully qualified domain name of the endpoint.

iBMDb2Settings_password :: Lens' IBMDb2Settings (Maybe Text)

Endpoint connection password.

iBMDb2Settings_port :: Lens' IBMDb2Settings (Maybe Int)

Endpoint TCP port. The default value is 50000.

iBMDb2Settings_username :: Lens' IBMDb2Settings (Maybe Text)

Endpoint connection user name.

iBMDb2Settings_secretsManagerAccessRoleArn :: Lens' IBMDb2Settings (Maybe Text)

The full Amazon Resource Name (ARN) of the IAM role that specifies DMS as the trusted entity and grants the required permissions to access the value in SecretsManagerSecret. The role must allow the iam:PassRole action. SecretsManagerSecret has the value of the Amazon Web Services Secrets Manager secret that allows access to the Db2 LUW endpoint.

You can specify one of two sets of values for these permissions. You can specify the values for this setting and SecretsManagerSecretId. Or you can specify clear-text values for UserName, Password, ServerName, and Port. You can't specify both. For more information on creating this SecretsManagerSecret and the SecretsManagerAccessRoleArn and SecretsManagerSecretId required to access it, see Using secrets to access Database Migration Service resources in the Database Migration Service User Guide.

iBMDb2Settings_maxKBytesPerRead :: Lens' IBMDb2Settings (Maybe Int)

Maximum number of bytes per read, as a NUMBER value. The default is 64 KB.

iBMDb2Settings_setDataCaptureChanges :: Lens' IBMDb2Settings (Maybe Bool)

Enables ongoing replication (CDC) as a BOOLEAN value. The default is true.

iBMDb2Settings_databaseName :: Lens' IBMDb2Settings (Maybe Text)

Database name for the endpoint.