Network.AWS.DMS.Types.KinesisSettings

Description

Documentation

data KinesisSettings

Provides information that describes an Amazon Kinesis Data Stream endpoint. This information includes the output format of records applied to the endpoint and details of transaction and control table data information.

See: newKinesisSettings smart constructor.

Constructors

KinesisSettings'

Fields

noHexPrefix :: Maybe Bool

Set this optional parameter to true to avoid adding a '0x' prefix to raw data in hexadecimal format. For example, by default, DMS adds a '0x' prefix to the LOB column type in hexadecimal format moving from an Oracle source to an Amazon Kinesis target. Use the NoHexPrefix endpoint setting to enable migration of RAW data type columns without adding the '0x' prefix.

includeNullAndEmpty :: Maybe Bool

Include NULL and empty columns for records migrated to the endpoint. The default is false.

messageFormat :: Maybe MessageFormatValue

The output format for the records created on the endpoint. The message format is JSON (default) or JSON_UNFORMATTED (a single line with no tab).

serviceAccessRoleArn :: Maybe Text

The Amazon Resource Name (ARN) for the IAM role that DMS uses to write to the Kinesis data stream. The role must allow the iam:PassRole action.

includeControlDetails :: Maybe Bool

Shows detailed control information for table definition, column definition, and table and column changes in the Kinesis message output. The default is false.

partitionIncludeSchemaTable :: Maybe Bool

Prefixes schema and table names to partition values, when the partition type is primary-key-type. Doing this increases data distribution among Kinesis shards. For example, suppose that a SysBench schema has thousands of tables and each table has only limited range for a primary key. In this case, the same primary key is sent from thousands of tables to the same shard, which causes throttling. The default is false.

streamArn :: Maybe Text

The Amazon Resource Name (ARN) for the Amazon Kinesis Data Streams endpoint.

includePartitionValue :: Maybe Bool

Shows the partition value within the Kinesis message output, unless the partition type is schema-table-type. The default is false.

includeTransactionDetails :: Maybe Bool

Provides detailed transaction information from the source database. This information includes a commit timestamp, a log position, and values for transaction_id, previous transaction_id, and transaction_record_id (the record offset within a transaction). The default is false.

includeTableAlterOperations :: Maybe Bool

Includes any data definition language (DDL) operations that change the table in the control data, such as rename-table, drop-table, add-column, drop-column, and rename-column. The default is false.

newKinesisSettings :: KinesisSettings

Create a value of KinesisSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:noHexPrefix:KinesisSettings', kinesisSettings_noHexPrefix - Set this optional parameter to true to avoid adding a '0x' prefix to raw data in hexadecimal format. For example, by default, DMS adds a '0x' prefix to the LOB column type in hexadecimal format moving from an Oracle source to an Amazon Kinesis target. Use the NoHexPrefix endpoint setting to enable migration of RAW data type columns without adding the '0x' prefix.

$sel:includeNullAndEmpty:KinesisSettings', kinesisSettings_includeNullAndEmpty - Include NULL and empty columns for records migrated to the endpoint. The default is false.

$sel:messageFormat:KinesisSettings', kinesisSettings_messageFormat - The output format for the records created on the endpoint. The message format is JSON (default) or JSON_UNFORMATTED (a single line with no tab).

$sel:serviceAccessRoleArn:KinesisSettings', kinesisSettings_serviceAccessRoleArn - The Amazon Resource Name (ARN) for the IAM role that DMS uses to write to the Kinesis data stream. The role must allow the iam:PassRole action.

$sel:includeControlDetails:KinesisSettings', kinesisSettings_includeControlDetails - Shows detailed control information for table definition, column definition, and table and column changes in the Kinesis message output. The default is false.

$sel:partitionIncludeSchemaTable:KinesisSettings', kinesisSettings_partitionIncludeSchemaTable - Prefixes schema and table names to partition values, when the partition type is primary-key-type. Doing this increases data distribution among Kinesis shards. For example, suppose that a SysBench schema has thousands of tables and each table has only limited range for a primary key. In this case, the same primary key is sent from thousands of tables to the same shard, which causes throttling. The default is false.

$sel:streamArn:KinesisSettings', kinesisSettings_streamArn - The Amazon Resource Name (ARN) for the Amazon Kinesis Data Streams endpoint.

$sel:includePartitionValue:KinesisSettings', kinesisSettings_includePartitionValue - Shows the partition value within the Kinesis message output, unless the partition type is schema-table-type. The default is false.

$sel:includeTransactionDetails:KinesisSettings', kinesisSettings_includeTransactionDetails - Provides detailed transaction information from the source database. This information includes a commit timestamp, a log position, and values for transaction_id, previous transaction_id, and transaction_record_id (the record offset within a transaction). The default is false.

$sel:includeTableAlterOperations:KinesisSettings', kinesisSettings_includeTableAlterOperations - Includes any data definition language (DDL) operations that change the table in the control data, such as rename-table, drop-table, add-column, drop-column, and rename-column. The default is false.

kinesisSettings_noHexPrefix :: Lens' KinesisSettings (Maybe Bool)

Set this optional parameter to true to avoid adding a '0x' prefix to raw data in hexadecimal format. For example, by default, DMS adds a '0x' prefix to the LOB column type in hexadecimal format moving from an Oracle source to an Amazon Kinesis target. Use the NoHexPrefix endpoint setting to enable migration of RAW data type columns without adding the '0x' prefix.

kinesisSettings_includeNullAndEmpty :: Lens' KinesisSettings (Maybe Bool)

Include NULL and empty columns for records migrated to the endpoint. The default is false.

kinesisSettings_messageFormat :: Lens' KinesisSettings (Maybe MessageFormatValue)

The output format for the records created on the endpoint. The message format is JSON (default) or JSON_UNFORMATTED (a single line with no tab).

kinesisSettings_serviceAccessRoleArn :: Lens' KinesisSettings (Maybe Text)

The Amazon Resource Name (ARN) for the IAM role that DMS uses to write to the Kinesis data stream. The role must allow the iam:PassRole action.

kinesisSettings_includeControlDetails :: Lens' KinesisSettings (Maybe Bool)

Shows detailed control information for table definition, column definition, and table and column changes in the Kinesis message output. The default is false.

kinesisSettings_partitionIncludeSchemaTable :: Lens' KinesisSettings (Maybe Bool)

Prefixes schema and table names to partition values, when the partition type is primary-key-type. Doing this increases data distribution among Kinesis shards. For example, suppose that a SysBench schema has thousands of tables and each table has only limited range for a primary key. In this case, the same primary key is sent from thousands of tables to the same shard, which causes throttling. The default is false.

kinesisSettings_streamArn :: Lens' KinesisSettings (Maybe Text)

The Amazon Resource Name (ARN) for the Amazon Kinesis Data Streams endpoint.

kinesisSettings_includePartitionValue :: Lens' KinesisSettings (Maybe Bool)

Shows the partition value within the Kinesis message output, unless the partition type is schema-table-type. The default is false.

kinesisSettings_includeTransactionDetails :: Lens' KinesisSettings (Maybe Bool)

Provides detailed transaction information from the source database. This information includes a commit timestamp, a log position, and values for transaction_id, previous transaction_id, and transaction_record_id (the record offset within a transaction). The default is false.

kinesisSettings_includeTableAlterOperations :: Lens' KinesisSettings (Maybe Bool)

Includes any data definition language (DDL) operations that change the table in the control data, such as rename-table, drop-table, add-column, drop-column, and rename-column. The default is false.