Network.AWS.DMS.Types.MicrosoftSQLServerSettings

Description

Documentation

data MicrosoftSQLServerSettings

Provides information that defines a Microsoft SQL Server endpoint.

See: newMicrosoftSQLServerSettings smart constructor.

Constructors

MicrosoftSQLServerSettings'

Fields

bcpPacketSize :: Maybe Int

The maximum size of the packets (in bytes) used to transfer data using BCP.

useBcpFullLoad :: Maybe Bool

Use this to attribute to transfer data for full-load operations using BCP. When the target table contains an identity column that does not exist in the source table, you must disable the use BCP for loading table option.

serverName :: Maybe Text

Fully qualified domain name of the endpoint.

querySingleAlwaysOnNode :: Maybe Bool

Cleans and recreates table metadata information on the replication instance when a mismatch occurs. An example is a situation where running an alter DDL statement on a table might result in different information about the table cached in the replication instance.

secretsManagerAccessRoleArn :: Maybe Text

The full Amazon Resource Name (ARN) of the IAM role that specifies DMS as the trusted entity and grants the required permissions to access the value in SecretsManagerSecret. The role must allow the iam:PassRole action. SecretsManagerSecret has the value of the Amazon Web Services Secrets Manager secret that allows access to the SQL Server endpoint.

You can specify one of two sets of values for these permissions. You can specify the values for this setting and SecretsManagerSecretId. Or you can specify clear-text values for UserName, Password, ServerName, and Port. You can't specify both. For more information on creating this SecretsManagerSecret and the SecretsManagerAccessRoleArn and SecretsManagerSecretId required to access it, see Using secrets to access Database Migration Service resources in the Database Migration Service User Guide.

username :: Maybe Text

Endpoint connection user name.

safeguardPolicy :: Maybe SafeguardPolicy

Use this attribute to minimize the need to access the backup log and enable DMS to prevent truncation using one of the following two methods.

Start transactions in the database: This is the default method. When this method is used, DMS prevents TLOG truncation by mimicking a transaction in the database. As long as such a transaction is open, changes that appear after the transaction started aren't truncated. If you need Microsoft Replication to be enabled in your database, then you must choose this method.

Exclusively use sp_repldone within a single task: When this method is used, DMS reads the changes and then uses sp_repldone to mark the TLOG transactions as ready for truncation. Although this method doesn't involve any transactional activities, it can only be used when Microsoft Replication isn't running. Also, when using this method, only one DMS task can access the database at any given time. Therefore, if you need to run parallel DMS tasks against the same database, use the default method.

password :: Maybe (Sensitive Text)

Endpoint connection password.

databaseName :: Maybe Text

Database name for the endpoint.

secretsManagerSecretId :: Maybe Text

The full ARN, partial ARN, or friendly name of the SecretsManagerSecret that contains the SQL Server endpoint connection details.

readBackupOnly :: Maybe Bool

When this attribute is set to Y, DMS only reads changes from transaction log backups and doesn't read from the active transaction log file during ongoing replication. Setting this parameter to Y enables you to control active transaction log file growth during full load and ongoing replication tasks. However, it can add some source latency to ongoing replication.

useThirdPartyBackupDevice :: Maybe Bool

When this attribute is set to Y, DMS processes third-party transaction log backups if they are created in native format.

controlTablesFileGroup :: Maybe Text

Specifies a file group for the DMS internal tables. When the replication task starts, all the internal DMS control tables (awsdms_ apply_exception, awsdms_apply, awsdms_changes) are created for the specified file group.

port :: Maybe Int

Endpoint TCP port.

newMicrosoftSQLServerSettings :: MicrosoftSQLServerSettings

Create a value of MicrosoftSQLServerSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bcpPacketSize:MicrosoftSQLServerSettings', microsoftSQLServerSettings_bcpPacketSize - The maximum size of the packets (in bytes) used to transfer data using BCP.

$sel:useBcpFullLoad:MicrosoftSQLServerSettings', microsoftSQLServerSettings_useBcpFullLoad - Use this to attribute to transfer data for full-load operations using BCP. When the target table contains an identity column that does not exist in the source table, you must disable the use BCP for loading table option.

$sel:serverName:MicrosoftSQLServerSettings', microsoftSQLServerSettings_serverName - Fully qualified domain name of the endpoint.

$sel:querySingleAlwaysOnNode:MicrosoftSQLServerSettings', microsoftSQLServerSettings_querySingleAlwaysOnNode - Cleans and recreates table metadata information on the replication instance when a mismatch occurs. An example is a situation where running an alter DDL statement on a table might result in different information about the table cached in the replication instance.

$sel:secretsManagerAccessRoleArn:MicrosoftSQLServerSettings', microsoftSQLServerSettings_secretsManagerAccessRoleArn - The full Amazon Resource Name (ARN) of the IAM role that specifies DMS as the trusted entity and grants the required permissions to access the value in SecretsManagerSecret. The role must allow the iam:PassRole action. SecretsManagerSecret has the value of the Amazon Web Services Secrets Manager secret that allows access to the SQL Server endpoint.

You can specify one of two sets of values for these permissions. You can specify the values for this setting and SecretsManagerSecretId. Or you can specify clear-text values for UserName, Password, ServerName, and Port. You can't specify both. For more information on creating this SecretsManagerSecret and the SecretsManagerAccessRoleArn and SecretsManagerSecretId required to access it, see Using secrets to access Database Migration Service resources in the Database Migration Service User Guide.

$sel:username:MicrosoftSQLServerSettings', microsoftSQLServerSettings_username - Endpoint connection user name.

$sel:safeguardPolicy:MicrosoftSQLServerSettings', microsoftSQLServerSettings_safeguardPolicy - Use this attribute to minimize the need to access the backup log and enable DMS to prevent truncation using one of the following two methods.

Start transactions in the database: This is the default method. When this method is used, DMS prevents TLOG truncation by mimicking a transaction in the database. As long as such a transaction is open, changes that appear after the transaction started aren't truncated. If you need Microsoft Replication to be enabled in your database, then you must choose this method.

Exclusively use sp_repldone within a single task: When this method is used, DMS reads the changes and then uses sp_repldone to mark the TLOG transactions as ready for truncation. Although this method doesn't involve any transactional activities, it can only be used when Microsoft Replication isn't running. Also, when using this method, only one DMS task can access the database at any given time. Therefore, if you need to run parallel DMS tasks against the same database, use the default method.

$sel:password:MicrosoftSQLServerSettings', microsoftSQLServerSettings_password - Endpoint connection password.

$sel:databaseName:MicrosoftSQLServerSettings', microsoftSQLServerSettings_databaseName - Database name for the endpoint.

$sel:secretsManagerSecretId:MicrosoftSQLServerSettings', microsoftSQLServerSettings_secretsManagerSecretId - The full ARN, partial ARN, or friendly name of the SecretsManagerSecret that contains the SQL Server endpoint connection details.

$sel:readBackupOnly:MicrosoftSQLServerSettings', microsoftSQLServerSettings_readBackupOnly - When this attribute is set to Y, DMS only reads changes from transaction log backups and doesn't read from the active transaction log file during ongoing replication. Setting this parameter to Y enables you to control active transaction log file growth during full load and ongoing replication tasks. However, it can add some source latency to ongoing replication.

$sel:useThirdPartyBackupDevice:MicrosoftSQLServerSettings', microsoftSQLServerSettings_useThirdPartyBackupDevice - When this attribute is set to Y, DMS processes third-party transaction log backups if they are created in native format.

$sel:controlTablesFileGroup:MicrosoftSQLServerSettings', microsoftSQLServerSettings_controlTablesFileGroup - Specifies a file group for the DMS internal tables. When the replication task starts, all the internal DMS control tables (awsdms_ apply_exception, awsdms_apply, awsdms_changes) are created for the specified file group.

$sel:port:MicrosoftSQLServerSettings', microsoftSQLServerSettings_port - Endpoint TCP port.

microsoftSQLServerSettings_bcpPacketSize :: Lens' MicrosoftSQLServerSettings (Maybe Int)

The maximum size of the packets (in bytes) used to transfer data using BCP.

microsoftSQLServerSettings_useBcpFullLoad :: Lens' MicrosoftSQLServerSettings (Maybe Bool)

Use this to attribute to transfer data for full-load operations using BCP. When the target table contains an identity column that does not exist in the source table, you must disable the use BCP for loading table option.

microsoftSQLServerSettings_serverName :: Lens' MicrosoftSQLServerSettings (Maybe Text)

Fully qualified domain name of the endpoint.

microsoftSQLServerSettings_querySingleAlwaysOnNode :: Lens' MicrosoftSQLServerSettings (Maybe Bool)

Cleans and recreates table metadata information on the replication instance when a mismatch occurs. An example is a situation where running an alter DDL statement on a table might result in different information about the table cached in the replication instance.

microsoftSQLServerSettings_secretsManagerAccessRoleArn :: Lens' MicrosoftSQLServerSettings (Maybe Text)

The full Amazon Resource Name (ARN) of the IAM role that specifies DMS as the trusted entity and grants the required permissions to access the value in SecretsManagerSecret. The role must allow the iam:PassRole action. SecretsManagerSecret has the value of the Amazon Web Services Secrets Manager secret that allows access to the SQL Server endpoint.

You can specify one of two sets of values for these permissions. You can specify the values for this setting and SecretsManagerSecretId. Or you can specify clear-text values for UserName, Password, ServerName, and Port. You can't specify both. For more information on creating this SecretsManagerSecret and the SecretsManagerAccessRoleArn and SecretsManagerSecretId required to access it, see Using secrets to access Database Migration Service resources in the Database Migration Service User Guide.

microsoftSQLServerSettings_username :: Lens' MicrosoftSQLServerSettings (Maybe Text)

Endpoint connection user name.

microsoftSQLServerSettings_safeguardPolicy :: Lens' MicrosoftSQLServerSettings (Maybe SafeguardPolicy)

Use this attribute to minimize the need to access the backup log and enable DMS to prevent truncation using one of the following two methods.

Start transactions in the database: This is the default method. When this method is used, DMS prevents TLOG truncation by mimicking a transaction in the database. As long as such a transaction is open, changes that appear after the transaction started aren't truncated. If you need Microsoft Replication to be enabled in your database, then you must choose this method.

Exclusively use sp_repldone within a single task: When this method is used, DMS reads the changes and then uses sp_repldone to mark the TLOG transactions as ready for truncation. Although this method doesn't involve any transactional activities, it can only be used when Microsoft Replication isn't running. Also, when using this method, only one DMS task can access the database at any given time. Therefore, if you need to run parallel DMS tasks against the same database, use the default method.

microsoftSQLServerSettings_password :: Lens' MicrosoftSQLServerSettings (Maybe Text)

Endpoint connection password.

microsoftSQLServerSettings_databaseName :: Lens' MicrosoftSQLServerSettings (Maybe Text)

Database name for the endpoint.

microsoftSQLServerSettings_secretsManagerSecretId :: Lens' MicrosoftSQLServerSettings (Maybe Text)

The full ARN, partial ARN, or friendly name of the SecretsManagerSecret that contains the SQL Server endpoint connection details.

microsoftSQLServerSettings_readBackupOnly :: Lens' MicrosoftSQLServerSettings (Maybe Bool)

When this attribute is set to Y, DMS only reads changes from transaction log backups and doesn't read from the active transaction log file during ongoing replication. Setting this parameter to Y enables you to control active transaction log file growth during full load and ongoing replication tasks. However, it can add some source latency to ongoing replication.

microsoftSQLServerSettings_useThirdPartyBackupDevice :: Lens' MicrosoftSQLServerSettings (Maybe Bool)

When this attribute is set to Y, DMS processes third-party transaction log backups if they are created in native format.

microsoftSQLServerSettings_controlTablesFileGroup :: Lens' MicrosoftSQLServerSettings (Maybe Text)

Specifies a file group for the DMS internal tables. When the replication task starts, all the internal DMS control tables (awsdms_ apply_exception, awsdms_apply, awsdms_changes) are created for the specified file group.

microsoftSQLServerSettings_port :: Lens' MicrosoftSQLServerSettings (Maybe Int)

Endpoint TCP port.