Network.AWS.DMS.Types.MongoDbSettings

Description

Documentation

data MongoDbSettings

Provides information that defines a MongoDB endpoint.

See: newMongoDbSettings smart constructor.

Constructors

MongoDbSettings'

Fields

serverName :: Maybe Text

The name of the server on the MongoDB source endpoint.

secretsManagerAccessRoleArn :: Maybe Text

The full Amazon Resource Name (ARN) of the IAM role that specifies DMS as the trusted entity and grants the required permissions to access the value in SecretsManagerSecret. The role must allow the iam:PassRole action. SecretsManagerSecret has the value of the Amazon Web Services Secrets Manager secret that allows access to the MongoDB endpoint.

You can specify one of two sets of values for these permissions. You can specify the values for this setting and SecretsManagerSecretId. Or you can specify clear-text values for UserName, Password, ServerName, and Port. You can't specify both. For more information on creating this SecretsManagerSecret and the SecretsManagerAccessRoleArn and SecretsManagerSecretId required to access it, see Using secrets to access Database Migration Service resources in the Database Migration Service User Guide.

authMechanism :: Maybe AuthMechanismValue

The authentication mechanism you use to access the MongoDB source endpoint.

For the default value, in MongoDB version 2.x, "default" is "mongodb_cr". For MongoDB version 3.x or later, "default" is "scram_sha_1". This setting isn't used when AuthType is set to "no".

username :: Maybe Text

The user name you use to access the MongoDB source endpoint.

kmsKeyId :: Maybe Text

The KMS key identifier that is used to encrypt the content on the replication instance. If you don't specify a value for the KmsKeyId parameter, then DMS uses your default encryption key. KMS creates the default encryption key for your Amazon Web Services account. Your Amazon Web Services account has a different default encryption key for each Amazon Web Services Region.

password :: Maybe (Sensitive Text)

The password for the user account you use to access the MongoDB source endpoint.

nestingLevel :: Maybe NestingLevelValue

Specifies either document or table mode.

Default value is "none". Specify "none" to use document mode. Specify "one" to use table mode.

databaseName :: Maybe Text

The database name on the MongoDB source endpoint.

docsToInvestigate :: Maybe Text

Indicates the number of documents to preview to determine the document organization. Use this setting when NestingLevel is set to "one".

Must be a positive value greater than 0. Default value is 1000.

authSource :: Maybe Text

The MongoDB database name. This setting isn't used when AuthType is set to "no".

The default is "admin".

secretsManagerSecretId :: Maybe Text

The full ARN, partial ARN, or friendly name of the SecretsManagerSecret that contains the MongoDB endpoint connection details.

extractDocId :: Maybe Text

Specifies the document ID. Use this setting when NestingLevel is set to "none".

Default value is "false".

authType :: Maybe AuthTypeValue

The authentication type you use to access the MongoDB source endpoint.

When when set to "no", user name and password parameters are not used and can be empty.

port :: Maybe Int

The port value for the MongoDB source endpoint.

newMongoDbSettings :: MongoDbSettings

Create a value of MongoDbSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serverName:MongoDbSettings', mongoDbSettings_serverName - The name of the server on the MongoDB source endpoint.

$sel:secretsManagerAccessRoleArn:MongoDbSettings', mongoDbSettings_secretsManagerAccessRoleArn - The full Amazon Resource Name (ARN) of the IAM role that specifies DMS as the trusted entity and grants the required permissions to access the value in SecretsManagerSecret. The role must allow the iam:PassRole action. SecretsManagerSecret has the value of the Amazon Web Services Secrets Manager secret that allows access to the MongoDB endpoint.

You can specify one of two sets of values for these permissions. You can specify the values for this setting and SecretsManagerSecretId. Or you can specify clear-text values for UserName, Password, ServerName, and Port. You can't specify both. For more information on creating this SecretsManagerSecret and the SecretsManagerAccessRoleArn and SecretsManagerSecretId required to access it, see Using secrets to access Database Migration Service resources in the Database Migration Service User Guide.

$sel:authMechanism:MongoDbSettings', mongoDbSettings_authMechanism - The authentication mechanism you use to access the MongoDB source endpoint.

For the default value, in MongoDB version 2.x, "default" is "mongodb_cr". For MongoDB version 3.x or later, "default" is "scram_sha_1". This setting isn't used when AuthType is set to "no".

$sel:username:MongoDbSettings', mongoDbSettings_username - The user name you use to access the MongoDB source endpoint.

$sel:kmsKeyId:MongoDbSettings', mongoDbSettings_kmsKeyId - The KMS key identifier that is used to encrypt the content on the replication instance. If you don't specify a value for the KmsKeyId parameter, then DMS uses your default encryption key. KMS creates the default encryption key for your Amazon Web Services account. Your Amazon Web Services account has a different default encryption key for each Amazon Web Services Region.

$sel:password:MongoDbSettings', mongoDbSettings_password - The password for the user account you use to access the MongoDB source endpoint.

$sel:nestingLevel:MongoDbSettings', mongoDbSettings_nestingLevel - Specifies either document or table mode.

Default value is "none". Specify "none" to use document mode. Specify "one" to use table mode.

$sel:databaseName:MongoDbSettings', mongoDbSettings_databaseName - The database name on the MongoDB source endpoint.

$sel:docsToInvestigate:MongoDbSettings', mongoDbSettings_docsToInvestigate - Indicates the number of documents to preview to determine the document organization. Use this setting when NestingLevel is set to "one".

Must be a positive value greater than 0. Default value is 1000.

$sel:authSource:MongoDbSettings', mongoDbSettings_authSource - The MongoDB database name. This setting isn't used when AuthType is set to "no".

The default is "admin".

$sel:secretsManagerSecretId:MongoDbSettings', mongoDbSettings_secretsManagerSecretId - The full ARN, partial ARN, or friendly name of the SecretsManagerSecret that contains the MongoDB endpoint connection details.

$sel:extractDocId:MongoDbSettings', mongoDbSettings_extractDocId - Specifies the document ID. Use this setting when NestingLevel is set to "none".

Default value is "false".

$sel:authType:MongoDbSettings', mongoDbSettings_authType - The authentication type you use to access the MongoDB source endpoint.

When when set to "no", user name and password parameters are not used and can be empty.

$sel:port:MongoDbSettings', mongoDbSettings_port - The port value for the MongoDB source endpoint.

mongoDbSettings_serverName :: Lens' MongoDbSettings (Maybe Text)

The name of the server on the MongoDB source endpoint.

mongoDbSettings_secretsManagerAccessRoleArn :: Lens' MongoDbSettings (Maybe Text)

The full Amazon Resource Name (ARN) of the IAM role that specifies DMS as the trusted entity and grants the required permissions to access the value in SecretsManagerSecret. The role must allow the iam:PassRole action. SecretsManagerSecret has the value of the Amazon Web Services Secrets Manager secret that allows access to the MongoDB endpoint.

You can specify one of two sets of values for these permissions. You can specify the values for this setting and SecretsManagerSecretId. Or you can specify clear-text values for UserName, Password, ServerName, and Port. You can't specify both. For more information on creating this SecretsManagerSecret and the SecretsManagerAccessRoleArn and SecretsManagerSecretId required to access it, see Using secrets to access Database Migration Service resources in the Database Migration Service User Guide.

mongoDbSettings_authMechanism :: Lens' MongoDbSettings (Maybe AuthMechanismValue)

The authentication mechanism you use to access the MongoDB source endpoint.

For the default value, in MongoDB version 2.x, "default" is "mongodb_cr". For MongoDB version 3.x or later, "default" is "scram_sha_1". This setting isn't used when AuthType is set to "no".

mongoDbSettings_username :: Lens' MongoDbSettings (Maybe Text)

The user name you use to access the MongoDB source endpoint.

mongoDbSettings_kmsKeyId :: Lens' MongoDbSettings (Maybe Text)

The KMS key identifier that is used to encrypt the content on the replication instance. If you don't specify a value for the KmsKeyId parameter, then DMS uses your default encryption key. KMS creates the default encryption key for your Amazon Web Services account. Your Amazon Web Services account has a different default encryption key for each Amazon Web Services Region.

mongoDbSettings_password :: Lens' MongoDbSettings (Maybe Text)

The password for the user account you use to access the MongoDB source endpoint.

mongoDbSettings_nestingLevel :: Lens' MongoDbSettings (Maybe NestingLevelValue)

Specifies either document or table mode.

Default value is "none". Specify "none" to use document mode. Specify "one" to use table mode.

mongoDbSettings_databaseName :: Lens' MongoDbSettings (Maybe Text)

The database name on the MongoDB source endpoint.

mongoDbSettings_docsToInvestigate :: Lens' MongoDbSettings (Maybe Text)

Indicates the number of documents to preview to determine the document organization. Use this setting when NestingLevel is set to "one".

Must be a positive value greater than 0. Default value is 1000.

mongoDbSettings_authSource :: Lens' MongoDbSettings (Maybe Text)

The MongoDB database name. This setting isn't used when AuthType is set to "no".

The default is "admin".

mongoDbSettings_secretsManagerSecretId :: Lens' MongoDbSettings (Maybe Text)

The full ARN, partial ARN, or friendly name of the SecretsManagerSecret that contains the MongoDB endpoint connection details.

mongoDbSettings_extractDocId :: Lens' MongoDbSettings (Maybe Text)

Specifies the document ID. Use this setting when NestingLevel is set to "none".

Default value is "false".

mongoDbSettings_authType :: Lens' MongoDbSettings (Maybe AuthTypeValue)

The authentication type you use to access the MongoDB source endpoint.

When when set to "no", user name and password parameters are not used and can be empty.

mongoDbSettings_port :: Lens' MongoDbSettings (Maybe Int)

The port value for the MongoDB source endpoint.