Network.AWS.DMS.Types.MySQLSettings

Description

Documentation

data MySQLSettings

Provides information that defines a MySQL endpoint.

See: newMySQLSettings smart constructor.

Constructors

MySQLSettings'

Fields

targetDbType :: Maybe TargetDbType

Specifies where to migrate source tables on the target, either to a single database or multiple databases.

Example: targetDbType=MULTIPLE_DATABASES

serverTimezone :: Maybe Text

Specifies the time zone for the source MySQL database.

Example: serverTimezone=US/Pacific;

Note: Do not enclose time zones in single quotes.

afterConnectScript :: Maybe Text

Specifies a script to run immediately after DMS connects to the endpoint. The migration task continues running regardless if the SQL statement succeeds or fails.

For this parameter, provide the code of the script itself, not the name of a file containing the script.

secretsManagerSecretId :: Maybe Text

The full ARN, partial ARN, or friendly name of the SecretsManagerSecret that contains the MySQL endpoint connection details.

serverName :: Maybe Text

Fully qualified domain name of the endpoint.

maxFileSize :: Maybe Int

Specifies the maximum size (in KB) of any .csv file used to transfer data to a MySQL-compatible database.

Example: maxFileSize=512

password :: Maybe (Sensitive Text)

Endpoint connection password.

eventsPollInterval :: Maybe Int

Specifies how often to check the binary log for new changes/events when the database is idle.

Example: eventsPollInterval=5;

In the example, DMS checks for changes in the binary logs every five seconds.

port :: Maybe Int

Endpoint TCP port.

username :: Maybe Text

Endpoint connection user name.

secretsManagerAccessRoleArn :: Maybe Text

The full Amazon Resource Name (ARN) of the IAM role that specifies DMS as the trusted entity and grants the required permissions to access the value in SecretsManagerSecret. The role must allow the iam:PassRole action. SecretsManagerSecret has the value of the Amazon Web Services Secrets Manager secret that allows access to the MySQL endpoint.

You can specify one of two sets of values for these permissions. You can specify the values for this setting and SecretsManagerSecretId. Or you can specify clear-text values for UserName, Password, ServerName, and Port. You can't specify both. For more information on creating this SecretsManagerSecret and the SecretsManagerAccessRoleArn and SecretsManagerSecretId required to access it, see Using secrets to access Database Migration Service resources in the Database Migration Service User Guide.

parallelLoadThreads :: Maybe Int

Improves performance when loading data into the MySQL-compatible target database. Specifies how many threads to use to load the data into the MySQL-compatible target database. Setting a large number of threads can have an adverse effect on database performance, because a separate connection is required for each thread.

Example: parallelLoadThreads=1

cleanSourceMetadataOnMismatch :: Maybe Bool

Adjusts the behavior of DMS when migrating from an SQL Server source database that is hosted as part of an Always On availability group cluster. If you need DMS to poll all the nodes in the Always On cluster for transaction backups, set this attribute to false.

databaseName :: Maybe Text

Database name for the endpoint. For a MySQL source or target endpoint, don't explicitly specify the database using the DatabaseName request parameter on either the CreateEndpoint or ModifyEndpoint API call. Specifying DatabaseName when you create or modify a MySQL endpoint replicates all the task tables to this single database. For MySQL endpoints, you specify the database only when you specify the schema in the table-mapping rules of the DMS task.

newMySQLSettings :: MySQLSettings

Create a value of MySQLSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetDbType:MySQLSettings', mySQLSettings_targetDbType - Specifies where to migrate source tables on the target, either to a single database or multiple databases.

Example: targetDbType=MULTIPLE_DATABASES

$sel:serverTimezone:MySQLSettings', mySQLSettings_serverTimezone - Specifies the time zone for the source MySQL database.

Example: serverTimezone=US/Pacific;

Note: Do not enclose time zones in single quotes.

$sel:afterConnectScript:MySQLSettings', mySQLSettings_afterConnectScript - Specifies a script to run immediately after DMS connects to the endpoint. The migration task continues running regardless if the SQL statement succeeds or fails.

For this parameter, provide the code of the script itself, not the name of a file containing the script.

$sel:secretsManagerSecretId:MySQLSettings', mySQLSettings_secretsManagerSecretId - The full ARN, partial ARN, or friendly name of the SecretsManagerSecret that contains the MySQL endpoint connection details.

$sel:serverName:MySQLSettings', mySQLSettings_serverName - Fully qualified domain name of the endpoint.

$sel:maxFileSize:MySQLSettings', mySQLSettings_maxFileSize - Specifies the maximum size (in KB) of any .csv file used to transfer data to a MySQL-compatible database.

Example: maxFileSize=512

$sel:password:MySQLSettings', mySQLSettings_password - Endpoint connection password.

$sel:eventsPollInterval:MySQLSettings', mySQLSettings_eventsPollInterval - Specifies how often to check the binary log for new changes/events when the database is idle.

Example: eventsPollInterval=5;

In the example, DMS checks for changes in the binary logs every five seconds.

$sel:port:MySQLSettings', mySQLSettings_port - Endpoint TCP port.

$sel:username:MySQLSettings', mySQLSettings_username - Endpoint connection user name.

$sel:secretsManagerAccessRoleArn:MySQLSettings', mySQLSettings_secretsManagerAccessRoleArn - The full Amazon Resource Name (ARN) of the IAM role that specifies DMS as the trusted entity and grants the required permissions to access the value in SecretsManagerSecret. The role must allow the iam:PassRole action. SecretsManagerSecret has the value of the Amazon Web Services Secrets Manager secret that allows access to the MySQL endpoint.

You can specify one of two sets of values for these permissions. You can specify the values for this setting and SecretsManagerSecretId. Or you can specify clear-text values for UserName, Password, ServerName, and Port. You can't specify both. For more information on creating this SecretsManagerSecret and the SecretsManagerAccessRoleArn and SecretsManagerSecretId required to access it, see Using secrets to access Database Migration Service resources in the Database Migration Service User Guide.

$sel:parallelLoadThreads:MySQLSettings', mySQLSettings_parallelLoadThreads - Improves performance when loading data into the MySQL-compatible target database. Specifies how many threads to use to load the data into the MySQL-compatible target database. Setting a large number of threads can have an adverse effect on database performance, because a separate connection is required for each thread.

Example: parallelLoadThreads=1

$sel:cleanSourceMetadataOnMismatch:MySQLSettings', mySQLSettings_cleanSourceMetadataOnMismatch - Adjusts the behavior of DMS when migrating from an SQL Server source database that is hosted as part of an Always On availability group cluster. If you need DMS to poll all the nodes in the Always On cluster for transaction backups, set this attribute to false.

$sel:databaseName:MySQLSettings', mySQLSettings_databaseName - Database name for the endpoint. For a MySQL source or target endpoint, don't explicitly specify the database using the DatabaseName request parameter on either the CreateEndpoint or ModifyEndpoint API call. Specifying DatabaseName when you create or modify a MySQL endpoint replicates all the task tables to this single database. For MySQL endpoints, you specify the database only when you specify the schema in the table-mapping rules of the DMS task.

mySQLSettings_targetDbType :: Lens' MySQLSettings (Maybe TargetDbType)

Specifies where to migrate source tables on the target, either to a single database or multiple databases.

Example: targetDbType=MULTIPLE_DATABASES

mySQLSettings_serverTimezone :: Lens' MySQLSettings (Maybe Text)

Specifies the time zone for the source MySQL database.

Example: serverTimezone=US/Pacific;

Note: Do not enclose time zones in single quotes.

mySQLSettings_afterConnectScript :: Lens' MySQLSettings (Maybe Text)

Specifies a script to run immediately after DMS connects to the endpoint. The migration task continues running regardless if the SQL statement succeeds or fails.

For this parameter, provide the code of the script itself, not the name of a file containing the script.

mySQLSettings_secretsManagerSecretId :: Lens' MySQLSettings (Maybe Text)

The full ARN, partial ARN, or friendly name of the SecretsManagerSecret that contains the MySQL endpoint connection details.

mySQLSettings_serverName :: Lens' MySQLSettings (Maybe Text)

Fully qualified domain name of the endpoint.

mySQLSettings_maxFileSize :: Lens' MySQLSettings (Maybe Int)

Specifies the maximum size (in KB) of any .csv file used to transfer data to a MySQL-compatible database.

Example: maxFileSize=512

mySQLSettings_password :: Lens' MySQLSettings (Maybe Text)

Endpoint connection password.

mySQLSettings_eventsPollInterval :: Lens' MySQLSettings (Maybe Int)

Specifies how often to check the binary log for new changes/events when the database is idle.

Example: eventsPollInterval=5;

In the example, DMS checks for changes in the binary logs every five seconds.

mySQLSettings_port :: Lens' MySQLSettings (Maybe Int)

Endpoint TCP port.

mySQLSettings_username :: Lens' MySQLSettings (Maybe Text)

Endpoint connection user name.

mySQLSettings_secretsManagerAccessRoleArn :: Lens' MySQLSettings (Maybe Text)

The full Amazon Resource Name (ARN) of the IAM role that specifies DMS as the trusted entity and grants the required permissions to access the value in SecretsManagerSecret. The role must allow the iam:PassRole action. SecretsManagerSecret has the value of the Amazon Web Services Secrets Manager secret that allows access to the MySQL endpoint.

You can specify one of two sets of values for these permissions. You can specify the values for this setting and SecretsManagerSecretId. Or you can specify clear-text values for UserName, Password, ServerName, and Port. You can't specify both. For more information on creating this SecretsManagerSecret and the SecretsManagerAccessRoleArn and SecretsManagerSecretId required to access it, see Using secrets to access Database Migration Service resources in the Database Migration Service User Guide.

mySQLSettings_parallelLoadThreads :: Lens' MySQLSettings (Maybe Int)

Improves performance when loading data into the MySQL-compatible target database. Specifies how many threads to use to load the data into the MySQL-compatible target database. Setting a large number of threads can have an adverse effect on database performance, because a separate connection is required for each thread.

Example: parallelLoadThreads=1

mySQLSettings_cleanSourceMetadataOnMismatch :: Lens' MySQLSettings (Maybe Bool)

Adjusts the behavior of DMS when migrating from an SQL Server source database that is hosted as part of an Always On availability group cluster. If you need DMS to poll all the nodes in the Always On cluster for transaction backups, set this attribute to false.

mySQLSettings_databaseName :: Lens' MySQLSettings (Maybe Text)

Database name for the endpoint. For a MySQL source or target endpoint, don't explicitly specify the database using the DatabaseName request parameter on either the CreateEndpoint or ModifyEndpoint API call. Specifying DatabaseName when you create or modify a MySQL endpoint replicates all the task tables to this single database. For MySQL endpoints, you specify the database only when you specify the schema in the table-mapping rules of the DMS task.