Network.AWS.DMS.Types.OrderableReplicationInstance

Description

Documentation

data OrderableReplicationInstance

In response to the DescribeOrderableReplicationInstances operation, this object describes an available replication instance. This description includes the replication instance's type, engine version, and allocated storage.

See: newOrderableReplicationInstance smart constructor.

Constructors

OrderableReplicationInstance'

Fields

engineVersion :: Maybe Text

The version of the replication engine.

minAllocatedStorage :: Maybe Int

The minimum amount of storage (in gigabytes) that can be allocated for the replication instance.

releaseStatus :: Maybe ReleaseStatusValues

The value returned when the specified EngineVersion of the replication instance is in Beta or test mode. This indicates some features might not work as expected.

DMS supports the ReleaseStatus parameter in versions 3.1.4 and later.

includedAllocatedStorage :: Maybe Int

The amount of storage (in gigabytes) that is allocated for the replication instance.

availabilityZones :: Maybe [Text]

List of Availability Zones for this replication instance.

maxAllocatedStorage :: Maybe Int

The minimum amount of storage (in gigabytes) that can be allocated for the replication instance.

replicationInstanceClass :: Maybe Text

The compute and memory capacity of the replication instance as defined for the specified replication instance class. For example to specify the instance class dms.c4.large, set this parameter to "dms.c4.large".

For more information on the settings and capacities for the available replication instance classes, see Selecting the right DMS replication instance for your migration.

defaultAllocatedStorage :: Maybe Int

The default amount of storage (in gigabytes) that is allocated for the replication instance.

storageType :: Maybe Text

The type of storage used by the replication instance.

newOrderableReplicationInstance :: OrderableReplicationInstance

Create a value of OrderableReplicationInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:engineVersion:OrderableReplicationInstance', orderableReplicationInstance_engineVersion - The version of the replication engine.

$sel:minAllocatedStorage:OrderableReplicationInstance', orderableReplicationInstance_minAllocatedStorage - The minimum amount of storage (in gigabytes) that can be allocated for the replication instance.

$sel:releaseStatus:OrderableReplicationInstance', orderableReplicationInstance_releaseStatus - The value returned when the specified EngineVersion of the replication instance is in Beta or test mode. This indicates some features might not work as expected.

DMS supports the ReleaseStatus parameter in versions 3.1.4 and later.

$sel:includedAllocatedStorage:OrderableReplicationInstance', orderableReplicationInstance_includedAllocatedStorage - The amount of storage (in gigabytes) that is allocated for the replication instance.

$sel:availabilityZones:OrderableReplicationInstance', orderableReplicationInstance_availabilityZones - List of Availability Zones for this replication instance.

$sel:maxAllocatedStorage:OrderableReplicationInstance', orderableReplicationInstance_maxAllocatedStorage - The minimum amount of storage (in gigabytes) that can be allocated for the replication instance.

$sel:replicationInstanceClass:OrderableReplicationInstance', orderableReplicationInstance_replicationInstanceClass - The compute and memory capacity of the replication instance as defined for the specified replication instance class. For example to specify the instance class dms.c4.large, set this parameter to "dms.c4.large".

For more information on the settings and capacities for the available replication instance classes, see Selecting the right DMS replication instance for your migration.

$sel:defaultAllocatedStorage:OrderableReplicationInstance', orderableReplicationInstance_defaultAllocatedStorage - The default amount of storage (in gigabytes) that is allocated for the replication instance.

$sel:storageType:OrderableReplicationInstance', orderableReplicationInstance_storageType - The type of storage used by the replication instance.

orderableReplicationInstance_engineVersion :: Lens' OrderableReplicationInstance (Maybe Text)

The version of the replication engine.

orderableReplicationInstance_minAllocatedStorage :: Lens' OrderableReplicationInstance (Maybe Int)

The minimum amount of storage (in gigabytes) that can be allocated for the replication instance.

orderableReplicationInstance_releaseStatus :: Lens' OrderableReplicationInstance (Maybe ReleaseStatusValues)

The value returned when the specified EngineVersion of the replication instance is in Beta or test mode. This indicates some features might not work as expected.

DMS supports the ReleaseStatus parameter in versions 3.1.4 and later.

orderableReplicationInstance_includedAllocatedStorage :: Lens' OrderableReplicationInstance (Maybe Int)

The amount of storage (in gigabytes) that is allocated for the replication instance.

orderableReplicationInstance_availabilityZones :: Lens' OrderableReplicationInstance (Maybe [Text])

List of Availability Zones for this replication instance.

orderableReplicationInstance_maxAllocatedStorage :: Lens' OrderableReplicationInstance (Maybe Int)

The minimum amount of storage (in gigabytes) that can be allocated for the replication instance.

orderableReplicationInstance_replicationInstanceClass :: Lens' OrderableReplicationInstance (Maybe Text)

The compute and memory capacity of the replication instance as defined for the specified replication instance class. For example to specify the instance class dms.c4.large, set this parameter to "dms.c4.large".

For more information on the settings and capacities for the available replication instance classes, see Selecting the right DMS replication instance for your migration.

orderableReplicationInstance_defaultAllocatedStorage :: Lens' OrderableReplicationInstance (Maybe Int)

The default amount of storage (in gigabytes) that is allocated for the replication instance.

orderableReplicationInstance_storageType :: Lens' OrderableReplicationInstance (Maybe Text)

The type of storage used by the replication instance.