Network.AWS.DMS.Types.PendingMaintenanceAction

Description

Documentation

data PendingMaintenanceAction

Describes a maintenance action pending for an DMS resource, including when and how it will be applied. This data type is a response element to the DescribePendingMaintenanceActions operation.

See: newPendingMaintenanceAction smart constructor.

Constructors

PendingMaintenanceAction'

Fields

forcedApplyDate :: Maybe POSIX

The date when the maintenance action will be automatically applied. The maintenance action is applied to the resource on this date regardless of the maintenance window for the resource. If this date is specified, any immediate opt-in requests are ignored.

optInStatus :: Maybe Text

The type of opt-in request that has been received for the resource.

autoAppliedAfterDate :: Maybe POSIX

The date of the maintenance window when the action is to be applied. The maintenance action is applied to the resource during its first maintenance window after this date. If this date is specified, any next-maintenance opt-in requests are ignored.

currentApplyDate :: Maybe POSIX

The effective date when the pending maintenance action will be applied to the resource. This date takes into account opt-in requests received from the ApplyPendingMaintenanceAction API operation, and also the AutoAppliedAfterDate and ForcedApplyDate parameter values. This value is blank if an opt-in request has not been received and nothing has been specified for AutoAppliedAfterDate or ForcedApplyDate.

action :: Maybe Text

The type of pending maintenance action that is available for the resource.

description :: Maybe Text

A description providing more detail about the maintenance action.

newPendingMaintenanceAction :: PendingMaintenanceAction

Create a value of PendingMaintenanceAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:forcedApplyDate:PendingMaintenanceAction', pendingMaintenanceAction_forcedApplyDate - The date when the maintenance action will be automatically applied. The maintenance action is applied to the resource on this date regardless of the maintenance window for the resource. If this date is specified, any immediate opt-in requests are ignored.

$sel:optInStatus:PendingMaintenanceAction', pendingMaintenanceAction_optInStatus - The type of opt-in request that has been received for the resource.

$sel:autoAppliedAfterDate:PendingMaintenanceAction', pendingMaintenanceAction_autoAppliedAfterDate - The date of the maintenance window when the action is to be applied. The maintenance action is applied to the resource during its first maintenance window after this date. If this date is specified, any next-maintenance opt-in requests are ignored.

$sel:currentApplyDate:PendingMaintenanceAction', pendingMaintenanceAction_currentApplyDate - The effective date when the pending maintenance action will be applied to the resource. This date takes into account opt-in requests received from the ApplyPendingMaintenanceAction API operation, and also the AutoAppliedAfterDate and ForcedApplyDate parameter values. This value is blank if an opt-in request has not been received and nothing has been specified for AutoAppliedAfterDate or ForcedApplyDate.

$sel:action:PendingMaintenanceAction', pendingMaintenanceAction_action - The type of pending maintenance action that is available for the resource.

$sel:description:PendingMaintenanceAction', pendingMaintenanceAction_description - A description providing more detail about the maintenance action.

pendingMaintenanceAction_forcedApplyDate :: Lens' PendingMaintenanceAction (Maybe UTCTime)

The date when the maintenance action will be automatically applied. The maintenance action is applied to the resource on this date regardless of the maintenance window for the resource. If this date is specified, any immediate opt-in requests are ignored.

pendingMaintenanceAction_optInStatus :: Lens' PendingMaintenanceAction (Maybe Text)

The type of opt-in request that has been received for the resource.

pendingMaintenanceAction_autoAppliedAfterDate :: Lens' PendingMaintenanceAction (Maybe UTCTime)

The date of the maintenance window when the action is to be applied. The maintenance action is applied to the resource during its first maintenance window after this date. If this date is specified, any next-maintenance opt-in requests are ignored.

pendingMaintenanceAction_currentApplyDate :: Lens' PendingMaintenanceAction (Maybe UTCTime)

The effective date when the pending maintenance action will be applied to the resource. This date takes into account opt-in requests received from the ApplyPendingMaintenanceAction API operation, and also the AutoAppliedAfterDate and ForcedApplyDate parameter values. This value is blank if an opt-in request has not been received and nothing has been specified for AutoAppliedAfterDate or ForcedApplyDate.

pendingMaintenanceAction_action :: Lens' PendingMaintenanceAction (Maybe Text)

The type of pending maintenance action that is available for the resource.

pendingMaintenanceAction_description :: Lens' PendingMaintenanceAction (Maybe Text)

A description providing more detail about the maintenance action.