Network.AWS.DMS.Types.PostgreSQLSettings

Description

Documentation

data PostgreSQLSettings

Provides information that defines a PostgreSQL endpoint.

See: newPostgreSQLSettings smart constructor.

Constructors

PostgreSQLSettings'

Fields

failTasksOnLobTruncation :: Maybe Bool

When set to true, this value causes a task to fail if the actual size of a LOB column is greater than the specified LobMaxSize.

If task is set to Limited LOB mode and this option is set to true, the task fails instead of truncating the LOB data.

executeTimeout :: Maybe Int

Sets the client statement timeout for the PostgreSQL instance, in seconds. The default value is 60 seconds.

Example: executeTimeout=100;

heartbeatEnable :: Maybe Bool

The write-ahead log (WAL) heartbeat feature mimics a dummy transaction. By doing this, it prevents idle logical replication slots from holding onto old WAL logs, which can result in storage full situations on the source. This heartbeat keeps restart_lsn moving and prevents storage full scenarios.

heartbeatSchema :: Maybe Text

Sets the schema in which the heartbeat artifacts are created.

pluginName :: Maybe PluginNameValue

Specifies the plugin to use to create a replication slot.

captureDdls :: Maybe Bool

To capture DDL events, DMS creates various artifacts in the PostgreSQL database when the task starts. You can later remove these artifacts.

If this value is set to N, you don't have to create tables or triggers on the source database.

slotName :: Maybe Text

Sets the name of a previously created logical replication slot for a change data capture (CDC) load of the PostgreSQL source instance.

When used with the CdcStartPosition request parameter for the DMS API , this attribute also makes it possible to use native CDC start points. DMS verifies that the specified logical replication slot exists before starting the CDC load task. It also verifies that the task was created with a valid setting of CdcStartPosition. If the specified slot doesn't exist or the task doesn't have a valid CdcStartPosition setting, DMS raises an error.

For more information about setting the CdcStartPosition request parameter, see Determining a CDC native start point in the Database Migration Service User Guide. For more information about using CdcStartPosition, see CreateReplicationTask, StartReplicationTask, and ModifyReplicationTask.

ddlArtifactsSchema :: Maybe Text

The schema in which the operational DDL database artifacts are created.

Example: ddlArtifactsSchema=xyzddlschema;

afterConnectScript :: Maybe Text

For use with change data capture (CDC) only, this attribute has DMS bypass foreign keys and user triggers to reduce the time it takes to bulk load data.

Example: afterConnectScript=SET session_replication_role='replica'

secretsManagerSecretId :: Maybe Text

The full ARN, partial ARN, or friendly name of the SecretsManagerSecret that contains the PostgreSQL endpoint connection details.

serverName :: Maybe Text

Fully qualified domain name of the endpoint.

maxFileSize :: Maybe Int

Specifies the maximum size (in KB) of any .csv file used to transfer data to PostgreSQL.

Example: maxFileSize=512

password :: Maybe (Sensitive Text)

Endpoint connection password.

heartbeatFrequency :: Maybe Int

Sets the WAL heartbeat frequency (in minutes).

port :: Maybe Int

Endpoint TCP port.

username :: Maybe Text

Endpoint connection user name.

secretsManagerAccessRoleArn :: Maybe Text

The full Amazon Resource Name (ARN) of the IAM role that specifies DMS as the trusted entity and grants the required permissions to access the value in SecretsManagerSecret. The role must allow the iam:PassRole action. SecretsManagerSecret has the value of the Amazon Web Services Secrets Manager secret that allows access to the PostgreSQL endpoint.

You can specify one of two sets of values for these permissions. You can specify the values for this setting and SecretsManagerSecretId. Or you can specify clear-text values for UserName, Password, ServerName, and Port. You can't specify both. For more information on creating this SecretsManagerSecret and the SecretsManagerAccessRoleArn and SecretsManagerSecretId required to access it, see Using secrets to access Database Migration Service resources in the Database Migration Service User Guide.

databaseName :: Maybe Text

Database name for the endpoint.

newPostgreSQLSettings :: PostgreSQLSettings

Create a value of PostgreSQLSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failTasksOnLobTruncation:PostgreSQLSettings', postgreSQLSettings_failTasksOnLobTruncation - When set to true, this value causes a task to fail if the actual size of a LOB column is greater than the specified LobMaxSize.

If task is set to Limited LOB mode and this option is set to true, the task fails instead of truncating the LOB data.

$sel:executeTimeout:PostgreSQLSettings', postgreSQLSettings_executeTimeout - Sets the client statement timeout for the PostgreSQL instance, in seconds. The default value is 60 seconds.

Example: executeTimeout=100;

$sel:heartbeatEnable:PostgreSQLSettings', postgreSQLSettings_heartbeatEnable - The write-ahead log (WAL) heartbeat feature mimics a dummy transaction. By doing this, it prevents idle logical replication slots from holding onto old WAL logs, which can result in storage full situations on the source. This heartbeat keeps restart_lsn moving and prevents storage full scenarios.

$sel:heartbeatSchema:PostgreSQLSettings', postgreSQLSettings_heartbeatSchema - Sets the schema in which the heartbeat artifacts are created.

$sel:pluginName:PostgreSQLSettings', postgreSQLSettings_pluginName - Specifies the plugin to use to create a replication slot.

$sel:captureDdls:PostgreSQLSettings', postgreSQLSettings_captureDdls - To capture DDL events, DMS creates various artifacts in the PostgreSQL database when the task starts. You can later remove these artifacts.

If this value is set to N, you don't have to create tables or triggers on the source database.

$sel:slotName:PostgreSQLSettings', postgreSQLSettings_slotName - Sets the name of a previously created logical replication slot for a change data capture (CDC) load of the PostgreSQL source instance.

When used with the CdcStartPosition request parameter for the DMS API , this attribute also makes it possible to use native CDC start points. DMS verifies that the specified logical replication slot exists before starting the CDC load task. It also verifies that the task was created with a valid setting of CdcStartPosition. If the specified slot doesn't exist or the task doesn't have a valid CdcStartPosition setting, DMS raises an error.

For more information about setting the CdcStartPosition request parameter, see Determining a CDC native start point in the Database Migration Service User Guide. For more information about using CdcStartPosition, see CreateReplicationTask, StartReplicationTask, and ModifyReplicationTask.

$sel:ddlArtifactsSchema:PostgreSQLSettings', postgreSQLSettings_ddlArtifactsSchema - The schema in which the operational DDL database artifacts are created.

Example: ddlArtifactsSchema=xyzddlschema;

$sel:afterConnectScript:PostgreSQLSettings', postgreSQLSettings_afterConnectScript - For use with change data capture (CDC) only, this attribute has DMS bypass foreign keys and user triggers to reduce the time it takes to bulk load data.

Example: afterConnectScript=SET session_replication_role='replica'

$sel:secretsManagerSecretId:PostgreSQLSettings', postgreSQLSettings_secretsManagerSecretId - The full ARN, partial ARN, or friendly name of the SecretsManagerSecret that contains the PostgreSQL endpoint connection details.

$sel:serverName:PostgreSQLSettings', postgreSQLSettings_serverName - Fully qualified domain name of the endpoint.

$sel:maxFileSize:PostgreSQLSettings', postgreSQLSettings_maxFileSize - Specifies the maximum size (in KB) of any .csv file used to transfer data to PostgreSQL.

Example: maxFileSize=512

$sel:password:PostgreSQLSettings', postgreSQLSettings_password - Endpoint connection password.

$sel:heartbeatFrequency:PostgreSQLSettings', postgreSQLSettings_heartbeatFrequency - Sets the WAL heartbeat frequency (in minutes).

$sel:port:PostgreSQLSettings', postgreSQLSettings_port - Endpoint TCP port.

$sel:username:PostgreSQLSettings', postgreSQLSettings_username - Endpoint connection user name.

$sel:secretsManagerAccessRoleArn:PostgreSQLSettings', postgreSQLSettings_secretsManagerAccessRoleArn - The full Amazon Resource Name (ARN) of the IAM role that specifies DMS as the trusted entity and grants the required permissions to access the value in SecretsManagerSecret. The role must allow the iam:PassRole action. SecretsManagerSecret has the value of the Amazon Web Services Secrets Manager secret that allows access to the PostgreSQL endpoint.

You can specify one of two sets of values for these permissions. You can specify the values for this setting and SecretsManagerSecretId. Or you can specify clear-text values for UserName, Password, ServerName, and Port. You can't specify both. For more information on creating this SecretsManagerSecret and the SecretsManagerAccessRoleArn and SecretsManagerSecretId required to access it, see Using secrets to access Database Migration Service resources in the Database Migration Service User Guide.

$sel:databaseName:PostgreSQLSettings', postgreSQLSettings_databaseName - Database name for the endpoint.

postgreSQLSettings_failTasksOnLobTruncation :: Lens' PostgreSQLSettings (Maybe Bool)

When set to true, this value causes a task to fail if the actual size of a LOB column is greater than the specified LobMaxSize.

If task is set to Limited LOB mode and this option is set to true, the task fails instead of truncating the LOB data.

postgreSQLSettings_executeTimeout :: Lens' PostgreSQLSettings (Maybe Int)

Sets the client statement timeout for the PostgreSQL instance, in seconds. The default value is 60 seconds.

Example: executeTimeout=100;

postgreSQLSettings_heartbeatEnable :: Lens' PostgreSQLSettings (Maybe Bool)

The write-ahead log (WAL) heartbeat feature mimics a dummy transaction. By doing this, it prevents idle logical replication slots from holding onto old WAL logs, which can result in storage full situations on the source. This heartbeat keeps restart_lsn moving and prevents storage full scenarios.

postgreSQLSettings_heartbeatSchema :: Lens' PostgreSQLSettings (Maybe Text)

Sets the schema in which the heartbeat artifacts are created.

postgreSQLSettings_pluginName :: Lens' PostgreSQLSettings (Maybe PluginNameValue)

Specifies the plugin to use to create a replication slot.

postgreSQLSettings_captureDdls :: Lens' PostgreSQLSettings (Maybe Bool)

To capture DDL events, DMS creates various artifacts in the PostgreSQL database when the task starts. You can later remove these artifacts.

If this value is set to N, you don't have to create tables or triggers on the source database.

postgreSQLSettings_slotName :: Lens' PostgreSQLSettings (Maybe Text)

Sets the name of a previously created logical replication slot for a change data capture (CDC) load of the PostgreSQL source instance.

When used with the CdcStartPosition request parameter for the DMS API , this attribute also makes it possible to use native CDC start points. DMS verifies that the specified logical replication slot exists before starting the CDC load task. It also verifies that the task was created with a valid setting of CdcStartPosition. If the specified slot doesn't exist or the task doesn't have a valid CdcStartPosition setting, DMS raises an error.

For more information about setting the CdcStartPosition request parameter, see Determining a CDC native start point in the Database Migration Service User Guide. For more information about using CdcStartPosition, see CreateReplicationTask, StartReplicationTask, and ModifyReplicationTask.

postgreSQLSettings_ddlArtifactsSchema :: Lens' PostgreSQLSettings (Maybe Text)

The schema in which the operational DDL database artifacts are created.

Example: ddlArtifactsSchema=xyzddlschema;

postgreSQLSettings_afterConnectScript :: Lens' PostgreSQLSettings (Maybe Text)

For use with change data capture (CDC) only, this attribute has DMS bypass foreign keys and user triggers to reduce the time it takes to bulk load data.

Example: afterConnectScript=SET session_replication_role='replica'

postgreSQLSettings_secretsManagerSecretId :: Lens' PostgreSQLSettings (Maybe Text)

The full ARN, partial ARN, or friendly name of the SecretsManagerSecret that contains the PostgreSQL endpoint connection details.

postgreSQLSettings_serverName :: Lens' PostgreSQLSettings (Maybe Text)

Fully qualified domain name of the endpoint.

postgreSQLSettings_maxFileSize :: Lens' PostgreSQLSettings (Maybe Int)

Specifies the maximum size (in KB) of any .csv file used to transfer data to PostgreSQL.

Example: maxFileSize=512

postgreSQLSettings_password :: Lens' PostgreSQLSettings (Maybe Text)

Endpoint connection password.

postgreSQLSettings_heartbeatFrequency :: Lens' PostgreSQLSettings (Maybe Int)

Sets the WAL heartbeat frequency (in minutes).

postgreSQLSettings_port :: Lens' PostgreSQLSettings (Maybe Int)

Endpoint TCP port.

postgreSQLSettings_username :: Lens' PostgreSQLSettings (Maybe Text)

Endpoint connection user name.

postgreSQLSettings_secretsManagerAccessRoleArn :: Lens' PostgreSQLSettings (Maybe Text)

The full Amazon Resource Name (ARN) of the IAM role that specifies DMS as the trusted entity and grants the required permissions to access the value in SecretsManagerSecret. The role must allow the iam:PassRole action. SecretsManagerSecret has the value of the Amazon Web Services Secrets Manager secret that allows access to the PostgreSQL endpoint.

You can specify one of two sets of values for these permissions. You can specify the values for this setting and SecretsManagerSecretId. Or you can specify clear-text values for UserName, Password, ServerName, and Port. You can't specify both. For more information on creating this SecretsManagerSecret and the SecretsManagerAccessRoleArn and SecretsManagerSecretId required to access it, see Using secrets to access Database Migration Service resources in the Database Migration Service User Guide.

postgreSQLSettings_databaseName :: Lens' PostgreSQLSettings (Maybe Text)

Database name for the endpoint.