Network.AWS.DMS.Types.RefreshSchemasStatus

Description

Documentation

data RefreshSchemasStatus

Provides information that describes status of a schema at an endpoint specified by the DescribeRefreshSchemaStatus operation.

See: newRefreshSchemasStatus smart constructor.

Constructors

RefreshSchemasStatus'

Fields

status :: Maybe RefreshSchemasStatusTypeValue

The status of the schema.

lastFailureMessage :: Maybe Text

The last failure message for the schema.

endpointArn :: Maybe Text

The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.

lastRefreshDate :: Maybe POSIX

The date the schema was last refreshed.

replicationInstanceArn :: Maybe Text

The Amazon Resource Name (ARN) of the replication instance.

newRefreshSchemasStatus :: RefreshSchemasStatus

Create a value of RefreshSchemasStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:RefreshSchemasStatus', refreshSchemasStatus_status - The status of the schema.

$sel:lastFailureMessage:RefreshSchemasStatus', refreshSchemasStatus_lastFailureMessage - The last failure message for the schema.

$sel:endpointArn:RefreshSchemasStatus', refreshSchemasStatus_endpointArn - The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.

$sel:lastRefreshDate:RefreshSchemasStatus', refreshSchemasStatus_lastRefreshDate - The date the schema was last refreshed.

$sel:replicationInstanceArn:RefreshSchemasStatus', refreshSchemasStatus_replicationInstanceArn - The Amazon Resource Name (ARN) of the replication instance.

refreshSchemasStatus_status :: Lens' RefreshSchemasStatus (Maybe RefreshSchemasStatusTypeValue)

The status of the schema.

refreshSchemasStatus_lastFailureMessage :: Lens' RefreshSchemasStatus (Maybe Text)

The last failure message for the schema.

refreshSchemasStatus_endpointArn :: Lens' RefreshSchemasStatus (Maybe Text)

The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.

refreshSchemasStatus_lastRefreshDate :: Lens' RefreshSchemasStatus (Maybe UTCTime)

The date the schema was last refreshed.

refreshSchemasStatus_replicationInstanceArn :: Lens' RefreshSchemasStatus (Maybe Text)

The Amazon Resource Name (ARN) of the replication instance.