Network.AWS.DMS.Types.ReplicationInstance

Description

Documentation

data ReplicationInstance

Provides information that defines a replication instance.

See: newReplicationInstance smart constructor.

Constructors

ReplicationInstance'

Fields

vpcSecurityGroups :: Maybe [VpcSecurityGroupMembership]

The VPC security group for the instance.

replicationInstancePrivateIpAddress :: Maybe Text

The private IP address of the replication instance.

freeUntil :: Maybe POSIX

The expiration date of the free replication instance that is part of the Free DMS program.

instanceCreateTime :: Maybe POSIX

The time the replication instance was created.

replicationSubnetGroup :: Maybe ReplicationSubnetGroup

The subnet group for the replication instance.

publiclyAccessible :: Maybe Bool

Specifies the accessibility options for the replication instance. A value of true represents an instance with a public IP address. A value of false represents an instance with a private IP address. The default value is true.

multiAZ :: Maybe Bool

Specifies whether the replication instance is a Multi-AZ deployment. You can't set the AvailabilityZone parameter if the Multi-AZ parameter is set to true.

kmsKeyId :: Maybe Text

An KMS key identifier that is used to encrypt the data on the replication instance.

If you don't specify a value for the KmsKeyId parameter, then DMS uses your default encryption key.

KMS creates the default encryption key for your Amazon Web Services account. Your Amazon Web Services account has a different default encryption key for each Amazon Web Services Region.

availabilityZone :: Maybe Text

The Availability Zone for the instance.

engineVersion :: Maybe Text

The engine version number of the replication instance.

If an engine version number is not specified when a replication instance is created, the default is the latest engine version available.

When modifying a major engine version of an instance, also set AllowMajorVersionUpgrade to true.

preferredMaintenanceWindow :: Maybe Text

The maintenance window times for the replication instance. Any pending upgrades to the replication instance are performed during this time.

replicationInstancePrivateIpAddresses :: Maybe [Text]

One or more private IP addresses for the replication instance.

replicationInstanceStatus :: Maybe Text

The status of the replication instance. The possible return values include:

"available"

"creating"

"deleted"

"deleting"

"failed"

"modifying"

"upgrading"

"rebooting"

"resetting-master-credentials"

"storage-full"

"incompatible-credentials"

"incompatible-network"

"maintenance"

replicationInstanceIdentifier :: Maybe Text

The replication instance identifier is a required parameter. This parameter is stored as a lowercase string.

Constraints:

Must contain 1-63 alphanumeric characters or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Example: myrepinstance

pendingModifiedValues :: Maybe ReplicationPendingModifiedValues

The pending modification values.

replicationInstanceClass :: Maybe Text

The compute and memory capacity of the replication instance as defined for the specified replication instance class. It is a required parameter, although a default value is pre-selected in the DMS console.

For more information on the settings and capacities for the available replication instance classes, see Selecting the right DMS replication instance for your migration.

replicationInstancePublicIpAddress :: Maybe Text

The public IP address of the replication instance.

dnsNameServers :: Maybe Text

The DNS name servers supported for the replication instance to access your on-premise source or target database.

replicationInstanceArn :: Maybe Text

The Amazon Resource Name (ARN) of the replication instance.

allocatedStorage :: Maybe Int

The amount of storage (in gigabytes) that is allocated for the replication instance.

replicationInstancePublicIpAddresses :: Maybe [Text]

One or more public IP addresses for the replication instance.

secondaryAvailabilityZone :: Maybe Text

The Availability Zone of the standby replication instance in a Multi-AZ deployment.

autoMinorVersionUpgrade :: Maybe Bool

Boolean value indicating if minor version upgrades will be automatically applied to the instance.

newReplicationInstance :: ReplicationInstance

Create a value of ReplicationInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpcSecurityGroups:ReplicationInstance', replicationInstance_vpcSecurityGroups - The VPC security group for the instance.

$sel:replicationInstancePrivateIpAddress:ReplicationInstance', replicationInstance_replicationInstancePrivateIpAddress - The private IP address of the replication instance.

$sel:freeUntil:ReplicationInstance', replicationInstance_freeUntil - The expiration date of the free replication instance that is part of the Free DMS program.

$sel:instanceCreateTime:ReplicationInstance', replicationInstance_instanceCreateTime - The time the replication instance was created.

$sel:replicationSubnetGroup:ReplicationInstance', replicationInstance_replicationSubnetGroup - The subnet group for the replication instance.

$sel:publiclyAccessible:ReplicationInstance', replicationInstance_publiclyAccessible - Specifies the accessibility options for the replication instance. A value of true represents an instance with a public IP address. A value of false represents an instance with a private IP address. The default value is true.

$sel:multiAZ:ReplicationInstance', replicationInstance_multiAZ - Specifies whether the replication instance is a Multi-AZ deployment. You can't set the AvailabilityZone parameter if the Multi-AZ parameter is set to true.

$sel:kmsKeyId:ReplicationInstance', replicationInstance_kmsKeyId - An KMS key identifier that is used to encrypt the data on the replication instance.

If you don't specify a value for the KmsKeyId parameter, then DMS uses your default encryption key.

KMS creates the default encryption key for your Amazon Web Services account. Your Amazon Web Services account has a different default encryption key for each Amazon Web Services Region.

$sel:availabilityZone:ReplicationInstance', replicationInstance_availabilityZone - The Availability Zone for the instance.

$sel:engineVersion:ReplicationInstance', replicationInstance_engineVersion - The engine version number of the replication instance.

If an engine version number is not specified when a replication instance is created, the default is the latest engine version available.

When modifying a major engine version of an instance, also set AllowMajorVersionUpgrade to true.

$sel:preferredMaintenanceWindow:ReplicationInstance', replicationInstance_preferredMaintenanceWindow - The maintenance window times for the replication instance. Any pending upgrades to the replication instance are performed during this time.

$sel:replicationInstancePrivateIpAddresses:ReplicationInstance', replicationInstance_replicationInstancePrivateIpAddresses - One or more private IP addresses for the replication instance.

$sel:replicationInstanceStatus:ReplicationInstance', replicationInstance_replicationInstanceStatus - The status of the replication instance. The possible return values include:

"available"

"creating"

"deleted"

"deleting"

"failed"

"modifying"

"upgrading"

"rebooting"

"resetting-master-credentials"

"storage-full"

"incompatible-credentials"

"incompatible-network"

"maintenance"

$sel:replicationInstanceIdentifier:ReplicationInstance', replicationInstance_replicationInstanceIdentifier - The replication instance identifier is a required parameter. This parameter is stored as a lowercase string.

Constraints:

Must contain 1-63 alphanumeric characters or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Example: myrepinstance

$sel:pendingModifiedValues:ReplicationInstance', replicationInstance_pendingModifiedValues - The pending modification values.

$sel:replicationInstanceClass:ReplicationInstance', replicationInstance_replicationInstanceClass - The compute and memory capacity of the replication instance as defined for the specified replication instance class. It is a required parameter, although a default value is pre-selected in the DMS console.

For more information on the settings and capacities for the available replication instance classes, see Selecting the right DMS replication instance for your migration.

$sel:replicationInstancePublicIpAddress:ReplicationInstance', replicationInstance_replicationInstancePublicIpAddress - The public IP address of the replication instance.

$sel:dnsNameServers:ReplicationInstance', replicationInstance_dnsNameServers - The DNS name servers supported for the replication instance to access your on-premise source or target database.

$sel:replicationInstanceArn:ReplicationInstance', replicationInstance_replicationInstanceArn - The Amazon Resource Name (ARN) of the replication instance.

$sel:allocatedStorage:ReplicationInstance', replicationInstance_allocatedStorage - The amount of storage (in gigabytes) that is allocated for the replication instance.

$sel:replicationInstancePublicIpAddresses:ReplicationInstance', replicationInstance_replicationInstancePublicIpAddresses - One or more public IP addresses for the replication instance.

$sel:secondaryAvailabilityZone:ReplicationInstance', replicationInstance_secondaryAvailabilityZone - The Availability Zone of the standby replication instance in a Multi-AZ deployment.

$sel:autoMinorVersionUpgrade:ReplicationInstance', replicationInstance_autoMinorVersionUpgrade - Boolean value indicating if minor version upgrades will be automatically applied to the instance.

replicationInstance_vpcSecurityGroups :: Lens' ReplicationInstance (Maybe [VpcSecurityGroupMembership])

The VPC security group for the instance.

replicationInstance_replicationInstancePrivateIpAddress :: Lens' ReplicationInstance (Maybe Text)

The private IP address of the replication instance.

replicationInstance_freeUntil :: Lens' ReplicationInstance (Maybe UTCTime)

The expiration date of the free replication instance that is part of the Free DMS program.

replicationInstance_instanceCreateTime :: Lens' ReplicationInstance (Maybe UTCTime)

The time the replication instance was created.

replicationInstance_replicationSubnetGroup :: Lens' ReplicationInstance (Maybe ReplicationSubnetGroup)

The subnet group for the replication instance.

replicationInstance_publiclyAccessible :: Lens' ReplicationInstance (Maybe Bool)

Specifies the accessibility options for the replication instance. A value of true represents an instance with a public IP address. A value of false represents an instance with a private IP address. The default value is true.

replicationInstance_multiAZ :: Lens' ReplicationInstance (Maybe Bool)

Specifies whether the replication instance is a Multi-AZ deployment. You can't set the AvailabilityZone parameter if the Multi-AZ parameter is set to true.

replicationInstance_kmsKeyId :: Lens' ReplicationInstance (Maybe Text)

An KMS key identifier that is used to encrypt the data on the replication instance.

If you don't specify a value for the KmsKeyId parameter, then DMS uses your default encryption key.

KMS creates the default encryption key for your Amazon Web Services account. Your Amazon Web Services account has a different default encryption key for each Amazon Web Services Region.

replicationInstance_availabilityZone :: Lens' ReplicationInstance (Maybe Text)

The Availability Zone for the instance.

replicationInstance_engineVersion :: Lens' ReplicationInstance (Maybe Text)

The engine version number of the replication instance.

If an engine version number is not specified when a replication instance is created, the default is the latest engine version available.

When modifying a major engine version of an instance, also set AllowMajorVersionUpgrade to true.

replicationInstance_preferredMaintenanceWindow :: Lens' ReplicationInstance (Maybe Text)

The maintenance window times for the replication instance. Any pending upgrades to the replication instance are performed during this time.

replicationInstance_replicationInstancePrivateIpAddresses :: Lens' ReplicationInstance (Maybe [Text])

One or more private IP addresses for the replication instance.

replicationInstance_replicationInstanceStatus :: Lens' ReplicationInstance (Maybe Text)

The status of the replication instance. The possible return values include:

"available"

"creating"

"deleted"

"deleting"

"failed"

"modifying"

"upgrading"

"rebooting"

"resetting-master-credentials"

"storage-full"

"incompatible-credentials"

"incompatible-network"

"maintenance"

replicationInstance_replicationInstanceIdentifier :: Lens' ReplicationInstance (Maybe Text)

The replication instance identifier is a required parameter. This parameter is stored as a lowercase string.

Constraints:

Must contain 1-63 alphanumeric characters or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Example: myrepinstance

replicationInstance_pendingModifiedValues :: Lens' ReplicationInstance (Maybe ReplicationPendingModifiedValues)

The pending modification values.

replicationInstance_replicationInstanceClass :: Lens' ReplicationInstance (Maybe Text)

The compute and memory capacity of the replication instance as defined for the specified replication instance class. It is a required parameter, although a default value is pre-selected in the DMS console.

For more information on the settings and capacities for the available replication instance classes, see Selecting the right DMS replication instance for your migration.

replicationInstance_replicationInstancePublicIpAddress :: Lens' ReplicationInstance (Maybe Text)

The public IP address of the replication instance.

replicationInstance_dnsNameServers :: Lens' ReplicationInstance (Maybe Text)

The DNS name servers supported for the replication instance to access your on-premise source or target database.

replicationInstance_replicationInstanceArn :: Lens' ReplicationInstance (Maybe Text)

The Amazon Resource Name (ARN) of the replication instance.

replicationInstance_allocatedStorage :: Lens' ReplicationInstance (Maybe Int)

The amount of storage (in gigabytes) that is allocated for the replication instance.

replicationInstance_replicationInstancePublicIpAddresses :: Lens' ReplicationInstance (Maybe [Text])

One or more public IP addresses for the replication instance.

replicationInstance_secondaryAvailabilityZone :: Lens' ReplicationInstance (Maybe Text)

The Availability Zone of the standby replication instance in a Multi-AZ deployment.

replicationInstance_autoMinorVersionUpgrade :: Lens' ReplicationInstance (Maybe Bool)

Boolean value indicating if minor version upgrades will be automatically applied to the instance.