Network.AWS.DMS.Types.ReplicationInstanceTaskLog

Description

Documentation

data ReplicationInstanceTaskLog

Contains metadata for a replication instance task log.

See: newReplicationInstanceTaskLog smart constructor.

Constructors

ReplicationInstanceTaskLog'

Fields

replicationTaskName :: Maybe Text

The name of the replication task.

replicationTaskArn :: Maybe Text

The Amazon Resource Name (ARN) of the replication task.

replicationInstanceTaskLogSize :: Maybe Integer

The size, in bytes, of the replication task log.

newReplicationInstanceTaskLog :: ReplicationInstanceTaskLog

Create a value of ReplicationInstanceTaskLog with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:replicationTaskName:ReplicationInstanceTaskLog', replicationInstanceTaskLog_replicationTaskName - The name of the replication task.

$sel:replicationTaskArn:ReplicationInstanceTaskLog', replicationInstanceTaskLog_replicationTaskArn - The Amazon Resource Name (ARN) of the replication task.

$sel:replicationInstanceTaskLogSize:ReplicationInstanceTaskLog', replicationInstanceTaskLog_replicationInstanceTaskLogSize - The size, in bytes, of the replication task log.

replicationInstanceTaskLog_replicationTaskName :: Lens' ReplicationInstanceTaskLog (Maybe Text)

The name of the replication task.

replicationInstanceTaskLog_replicationTaskArn :: Lens' ReplicationInstanceTaskLog (Maybe Text)

The Amazon Resource Name (ARN) of the replication task.

replicationInstanceTaskLog_replicationInstanceTaskLogSize :: Lens' ReplicationInstanceTaskLog (Maybe Integer)

The size, in bytes, of the replication task log.