Network.AWS.DMS.Types.ReplicationSubnetGroup

Description

Documentation

data ReplicationSubnetGroup

Describes a subnet group in response to a request by the DescribeReplicationSubnetGroups operation.

See: newReplicationSubnetGroup smart constructor.

Constructors

ReplicationSubnetGroup'

Fields

replicationSubnetGroupIdentifier :: Maybe Text

The identifier of the replication instance subnet group.

subnetGroupStatus :: Maybe Text

The status of the subnet group.

replicationSubnetGroupDescription :: Maybe Text

A description for the replication subnet group.

vpcId :: Maybe Text

The ID of the VPC.

subnets :: Maybe [Subnet]

The subnets that are in the subnet group.

newReplicationSubnetGroup :: ReplicationSubnetGroup

Create a value of ReplicationSubnetGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:replicationSubnetGroupIdentifier:ReplicationSubnetGroup', replicationSubnetGroup_replicationSubnetGroupIdentifier - The identifier of the replication instance subnet group.

$sel:subnetGroupStatus:ReplicationSubnetGroup', replicationSubnetGroup_subnetGroupStatus - The status of the subnet group.

$sel:replicationSubnetGroupDescription:ReplicationSubnetGroup', replicationSubnetGroup_replicationSubnetGroupDescription - A description for the replication subnet group.

$sel:vpcId:ReplicationSubnetGroup', replicationSubnetGroup_vpcId - The ID of the VPC.

$sel:subnets:ReplicationSubnetGroup', replicationSubnetGroup_subnets - The subnets that are in the subnet group.

replicationSubnetGroup_replicationSubnetGroupIdentifier :: Lens' ReplicationSubnetGroup (Maybe Text)

The identifier of the replication instance subnet group.

replicationSubnetGroup_subnetGroupStatus :: Lens' ReplicationSubnetGroup (Maybe Text)

The status of the subnet group.

replicationSubnetGroup_replicationSubnetGroupDescription :: Lens' ReplicationSubnetGroup (Maybe Text)

A description for the replication subnet group.

replicationSubnetGroup_vpcId :: Lens' ReplicationSubnetGroup (Maybe Text)

The ID of the VPC.

replicationSubnetGroup_subnets :: Lens' ReplicationSubnetGroup (Maybe [Subnet])

The subnets that are in the subnet group.