Network.AWS.DMS.Types.ReplicationTask

Description

Documentation

data ReplicationTask

Provides information that describes a replication task created by the CreateReplicationTask operation.

See: newReplicationTask smart constructor.

Constructors

ReplicationTask'

Fields

replicationTaskSettings :: Maybe Text

The settings for the replication task.

status :: Maybe Text

The status of the replication task. This response parameter can return one of the following values:

"moving" – The task is being moved in response to running the MoveReplicationTask operation.

"creating" – The task is being created in response to running the CreateReplicationTask operation.

"deleting" – The task is being deleted in response to running the DeleteReplicationTask operation.

"failed" – The task failed to successfully complete the database migration in response to running the StartReplicationTask operation.

"failed-move" – The task failed to move in response to running the MoveReplicationTask operation.

"modifying" – The task definition is being modified in response to running the ModifyReplicationTask operation.

"ready" – The task is in a ready state where it can respond to other task operations, such as StartReplicationTask or DeleteReplicationTask .

"running" – The task is performing a database migration in response to running the StartReplicationTask operation.

"starting" – The task is preparing to perform a database migration in response to running the StartReplicationTask operation.

"stopped" – The task has stopped in response to running the StopReplicationTask operation.

"stopping" – The task is preparing to stop in response to running the StopReplicationTask operation.

"testing" – The database migration specified for this task is being tested in response to running either the StartReplicationTaskAssessmentRun or the StartReplicationTaskAssessment operation.

StartReplicationTaskAssessmentRun is an improved premigration task assessment operation. The StartReplicationTaskAssessment operation assesses data type compatibility only between the source and target database of a given migration task. In contrast, StartReplicationTaskAssessmentRun enables you to specify a variety of premigration task assessments in addition to data type compatibility. These assessments include ones for the validity of primary key definitions and likely issues with database migration performance, among others.

stopReason :: Maybe Text

The reason the replication task was stopped. This response parameter can return one of the following values:

"STOP_REASON_FULL_LOAD_COMPLETED" – Full-load migration completed.

"STOP_REASON_CACHED_CHANGES_APPLIED" – Change data capture (CDC) load completed.

"STOP_REASON_CACHED_CHANGES_NOT_APPLIED" – In a full-load and CDC migration, the full load stopped as specified before starting the CDC migration.

"STOP_REASON_SERVER_TIME" – The migration stopped at the specified server time.

targetEndpointArn :: Maybe Text

The ARN that uniquely identifies the endpoint.

replicationTaskIdentifier :: Maybe Text

The user-assigned replication task identifier or name.

Constraints:

Must contain 1-255 alphanumeric characters or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

cdcStartPosition :: Maybe Text

Indicates when you want a change data capture (CDC) operation to start. Use either CdcStartPosition or CdcStartTime to specify when you want the CDC operation to start. Specifying both values results in an error.

The value can be in date, checkpoint, or LSN/SCN format.

Date Example: --cdc-start-position “2018-03-08T12:12:12”

Checkpoint Example: --cdc-start-position "checkpoint:V1mysql-bin-changelog.157832:1975:-1:2002:677883278264080:mysql-bin-changelog.157832:187600#93"

LSN Example: --cdc-start-position “mysql-bin-changelog.000024:373”

replicationTaskStartDate :: Maybe POSIX

The date the replication task is scheduled to start.

sourceEndpointArn :: Maybe Text

The Amazon Resource Name (ARN) that uniquely identifies the endpoint.

recoveryCheckpoint :: Maybe Text

Indicates the last checkpoint that occurred during a change data capture (CDC) operation. You can provide this value to the CdcStartPosition parameter to start a CDC operation that begins at that checkpoint.

tableMappings :: Maybe Text

Table mappings specified in the task.

targetReplicationInstanceArn :: Maybe Text

The ARN of the replication instance to which this task is moved in response to running the MoveReplicationTask operation. Otherwise, this response parameter isn't a member of the ReplicationTask object.

replicationTaskCreationDate :: Maybe POSIX

The date the replication task was created.

migrationType :: Maybe MigrationTypeValue

The type of migration.

replicationTaskArn :: Maybe Text

The Amazon Resource Name (ARN) of the replication task.

taskData :: Maybe Text

Supplemental information that the task requires to migrate the data for certain source and target endpoints. For more information, see Specifying Supplemental Data for Task Settings in the Database Migration Service User Guide.

cdcStopPosition :: Maybe Text

Indicates when you want a change data capture (CDC) operation to stop. The value can be either server time or commit time.

Server time example: --cdc-stop-position “server_time:2018-02-09T12:12:12”

Commit time example: --cdc-stop-position “commit_time: 2018-02-09T12:12:12 “

replicationTaskStats :: Maybe ReplicationTaskStats

The statistics for the task, including elapsed time, tables loaded, and table errors.

replicationInstanceArn :: Maybe Text

The ARN of the replication instance.

lastFailureMessage :: Maybe Text

The last error (failure) message generated for the replication task.

newReplicationTask :: ReplicationTask

Create a value of ReplicationTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:replicationTaskSettings:ReplicationTask', replicationTask_replicationTaskSettings - The settings for the replication task.

$sel:status:ReplicationTask', replicationTask_status - The status of the replication task. This response parameter can return one of the following values:

"moving" – The task is being moved in response to running the MoveReplicationTask operation.

"creating" – The task is being created in response to running the CreateReplicationTask operation.

"deleting" – The task is being deleted in response to running the DeleteReplicationTask operation.

"failed" – The task failed to successfully complete the database migration in response to running the StartReplicationTask operation.

"failed-move" – The task failed to move in response to running the MoveReplicationTask operation.

"modifying" – The task definition is being modified in response to running the ModifyReplicationTask operation.

"ready" – The task is in a ready state where it can respond to other task operations, such as StartReplicationTask or DeleteReplicationTask .

"running" – The task is performing a database migration in response to running the StartReplicationTask operation.

"starting" – The task is preparing to perform a database migration in response to running the StartReplicationTask operation.

"stopped" – The task has stopped in response to running the StopReplicationTask operation.

"stopping" – The task is preparing to stop in response to running the StopReplicationTask operation.

"testing" – The database migration specified for this task is being tested in response to running either the StartReplicationTaskAssessmentRun or the StartReplicationTaskAssessment operation.

StartReplicationTaskAssessmentRun is an improved premigration task assessment operation. The StartReplicationTaskAssessment operation assesses data type compatibility only between the source and target database of a given migration task. In contrast, StartReplicationTaskAssessmentRun enables you to specify a variety of premigration task assessments in addition to data type compatibility. These assessments include ones for the validity of primary key definitions and likely issues with database migration performance, among others.

$sel:stopReason:ReplicationTask', replicationTask_stopReason - The reason the replication task was stopped. This response parameter can return one of the following values:

"STOP_REASON_FULL_LOAD_COMPLETED" – Full-load migration completed.

"STOP_REASON_CACHED_CHANGES_APPLIED" – Change data capture (CDC) load completed.

"STOP_REASON_CACHED_CHANGES_NOT_APPLIED" – In a full-load and CDC migration, the full load stopped as specified before starting the CDC migration.

"STOP_REASON_SERVER_TIME" – The migration stopped at the specified server time.

$sel:targetEndpointArn:ReplicationTask', replicationTask_targetEndpointArn - The ARN that uniquely identifies the endpoint.

$sel:replicationTaskIdentifier:ReplicationTask', replicationTask_replicationTaskIdentifier - The user-assigned replication task identifier or name.

Constraints:

Must contain 1-255 alphanumeric characters or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

$sel:cdcStartPosition:ReplicationTask', replicationTask_cdcStartPosition - Indicates when you want a change data capture (CDC) operation to start. Use either CdcStartPosition or CdcStartTime to specify when you want the CDC operation to start. Specifying both values results in an error.

The value can be in date, checkpoint, or LSN/SCN format.

Date Example: --cdc-start-position “2018-03-08T12:12:12”

Checkpoint Example: --cdc-start-position "checkpoint:V1mysql-bin-changelog.157832:1975:-1:2002:677883278264080:mysql-bin-changelog.157832:187600#93"

LSN Example: --cdc-start-position “mysql-bin-changelog.000024:373”

$sel:replicationTaskStartDate:ReplicationTask', replicationTask_replicationTaskStartDate - The date the replication task is scheduled to start.

$sel:sourceEndpointArn:ReplicationTask', replicationTask_sourceEndpointArn - The Amazon Resource Name (ARN) that uniquely identifies the endpoint.

$sel:recoveryCheckpoint:ReplicationTask', replicationTask_recoveryCheckpoint - Indicates the last checkpoint that occurred during a change data capture (CDC) operation. You can provide this value to the CdcStartPosition parameter to start a CDC operation that begins at that checkpoint.

$sel:tableMappings:ReplicationTask', replicationTask_tableMappings - Table mappings specified in the task.

$sel:targetReplicationInstanceArn:ReplicationTask', replicationTask_targetReplicationInstanceArn - The ARN of the replication instance to which this task is moved in response to running the MoveReplicationTask operation. Otherwise, this response parameter isn't a member of the ReplicationTask object.

$sel:replicationTaskCreationDate:ReplicationTask', replicationTask_replicationTaskCreationDate - The date the replication task was created.

$sel:migrationType:ReplicationTask', replicationTask_migrationType - The type of migration.

$sel:replicationTaskArn:ReplicationTask', replicationTask_replicationTaskArn - The Amazon Resource Name (ARN) of the replication task.

$sel:taskData:ReplicationTask', replicationTask_taskData - Supplemental information that the task requires to migrate the data for certain source and target endpoints. For more information, see Specifying Supplemental Data for Task Settings in the Database Migration Service User Guide.

$sel:cdcStopPosition:ReplicationTask', replicationTask_cdcStopPosition - Indicates when you want a change data capture (CDC) operation to stop. The value can be either server time or commit time.

Server time example: --cdc-stop-position “server_time:2018-02-09T12:12:12”

Commit time example: --cdc-stop-position “commit_time: 2018-02-09T12:12:12 “

$sel:replicationTaskStats:ReplicationTask', replicationTask_replicationTaskStats - The statistics for the task, including elapsed time, tables loaded, and table errors.

$sel:replicationInstanceArn:ReplicationTask', replicationTask_replicationInstanceArn - The ARN of the replication instance.

$sel:lastFailureMessage:ReplicationTask', replicationTask_lastFailureMessage - The last error (failure) message generated for the replication task.

replicationTask_replicationTaskSettings :: Lens' ReplicationTask (Maybe Text)

The settings for the replication task.

replicationTask_status :: Lens' ReplicationTask (Maybe Text)

The status of the replication task. This response parameter can return one of the following values:

"moving" – The task is being moved in response to running the MoveReplicationTask operation.

"creating" – The task is being created in response to running the CreateReplicationTask operation.

"deleting" – The task is being deleted in response to running the DeleteReplicationTask operation.

"failed" – The task failed to successfully complete the database migration in response to running the StartReplicationTask operation.

"failed-move" – The task failed to move in response to running the MoveReplicationTask operation.

"modifying" – The task definition is being modified in response to running the ModifyReplicationTask operation.

"ready" – The task is in a ready state where it can respond to other task operations, such as StartReplicationTask or DeleteReplicationTask .

"running" – The task is performing a database migration in response to running the StartReplicationTask operation.

"starting" – The task is preparing to perform a database migration in response to running the StartReplicationTask operation.

"stopped" – The task has stopped in response to running the StopReplicationTask operation.

"stopping" – The task is preparing to stop in response to running the StopReplicationTask operation.

"testing" – The database migration specified for this task is being tested in response to running either the StartReplicationTaskAssessmentRun or the StartReplicationTaskAssessment operation.

StartReplicationTaskAssessmentRun is an improved premigration task assessment operation. The StartReplicationTaskAssessment operation assesses data type compatibility only between the source and target database of a given migration task. In contrast, StartReplicationTaskAssessmentRun enables you to specify a variety of premigration task assessments in addition to data type compatibility. These assessments include ones for the validity of primary key definitions and likely issues with database migration performance, among others.

replicationTask_stopReason :: Lens' ReplicationTask (Maybe Text)

The reason the replication task was stopped. This response parameter can return one of the following values:

"STOP_REASON_FULL_LOAD_COMPLETED" – Full-load migration completed.

"STOP_REASON_CACHED_CHANGES_APPLIED" – Change data capture (CDC) load completed.

"STOP_REASON_CACHED_CHANGES_NOT_APPLIED" – In a full-load and CDC migration, the full load stopped as specified before starting the CDC migration.

"STOP_REASON_SERVER_TIME" – The migration stopped at the specified server time.

replicationTask_targetEndpointArn :: Lens' ReplicationTask (Maybe Text)

The ARN that uniquely identifies the endpoint.

replicationTask_replicationTaskIdentifier :: Lens' ReplicationTask (Maybe Text)

The user-assigned replication task identifier or name.

Constraints:

Must contain 1-255 alphanumeric characters or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

replicationTask_cdcStartPosition :: Lens' ReplicationTask (Maybe Text)

Indicates when you want a change data capture (CDC) operation to start. Use either CdcStartPosition or CdcStartTime to specify when you want the CDC operation to start. Specifying both values results in an error.

The value can be in date, checkpoint, or LSN/SCN format.

Date Example: --cdc-start-position “2018-03-08T12:12:12”

Checkpoint Example: --cdc-start-position "checkpoint:V1mysql-bin-changelog.157832:1975:-1:2002:677883278264080:mysql-bin-changelog.157832:187600#93"

LSN Example: --cdc-start-position “mysql-bin-changelog.000024:373”

replicationTask_replicationTaskStartDate :: Lens' ReplicationTask (Maybe UTCTime)

The date the replication task is scheduled to start.

replicationTask_sourceEndpointArn :: Lens' ReplicationTask (Maybe Text)

The Amazon Resource Name (ARN) that uniquely identifies the endpoint.

replicationTask_recoveryCheckpoint :: Lens' ReplicationTask (Maybe Text)

Indicates the last checkpoint that occurred during a change data capture (CDC) operation. You can provide this value to the CdcStartPosition parameter to start a CDC operation that begins at that checkpoint.

replicationTask_tableMappings :: Lens' ReplicationTask (Maybe Text)

Table mappings specified in the task.

replicationTask_targetReplicationInstanceArn :: Lens' ReplicationTask (Maybe Text)

The ARN of the replication instance to which this task is moved in response to running the MoveReplicationTask operation. Otherwise, this response parameter isn't a member of the ReplicationTask object.

replicationTask_replicationTaskCreationDate :: Lens' ReplicationTask (Maybe UTCTime)

The date the replication task was created.

replicationTask_migrationType :: Lens' ReplicationTask (Maybe MigrationTypeValue)

The type of migration.

replicationTask_replicationTaskArn :: Lens' ReplicationTask (Maybe Text)

The Amazon Resource Name (ARN) of the replication task.

replicationTask_taskData :: Lens' ReplicationTask (Maybe Text)

Supplemental information that the task requires to migrate the data for certain source and target endpoints. For more information, see Specifying Supplemental Data for Task Settings in the Database Migration Service User Guide.

replicationTask_cdcStopPosition :: Lens' ReplicationTask (Maybe Text)

Indicates when you want a change data capture (CDC) operation to stop. The value can be either server time or commit time.

Server time example: --cdc-stop-position “server_time:2018-02-09T12:12:12”

Commit time example: --cdc-stop-position “commit_time: 2018-02-09T12:12:12 “

replicationTask_replicationTaskStats :: Lens' ReplicationTask (Maybe ReplicationTaskStats)

The statistics for the task, including elapsed time, tables loaded, and table errors.

replicationTask_replicationInstanceArn :: Lens' ReplicationTask (Maybe Text)

The ARN of the replication instance.

replicationTask_lastFailureMessage :: Lens' ReplicationTask (Maybe Text)

The last error (failure) message generated for the replication task.