Network.AWS.DMS.Types.ReplicationTaskAssessmentResult

Description

Documentation

data ReplicationTaskAssessmentResult

The task assessment report in JSON format.

See: newReplicationTaskAssessmentResult smart constructor.

Constructors

ReplicationTaskAssessmentResult'

Fields

assessmentResults :: Maybe Text

The task assessment results in JSON format.

The response object only contains this field if you provide DescribeReplicationTaskAssessmentResultsMessage$ReplicationTaskArn in the request.

assessmentResultsFile :: Maybe Text

The file containing the results of the task assessment.

replicationTaskIdentifier :: Maybe Text

The replication task identifier of the task on which the task assessment was run.

assessmentStatus :: Maybe Text

The status of the task assessment.

s3ObjectUrl :: Maybe Text

The URL of the S3 object containing the task assessment results.

The response object only contains this field if you provide DescribeReplicationTaskAssessmentResultsMessage$ReplicationTaskArn in the request.

replicationTaskLastAssessmentDate :: Maybe POSIX

The date the task assessment was completed.

replicationTaskArn :: Maybe Text

The Amazon Resource Name (ARN) of the replication task.

newReplicationTaskAssessmentResult :: ReplicationTaskAssessmentResult

Create a value of ReplicationTaskAssessmentResult with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:assessmentResults:ReplicationTaskAssessmentResult', replicationTaskAssessmentResult_assessmentResults - The task assessment results in JSON format.

The response object only contains this field if you provide DescribeReplicationTaskAssessmentResultsMessage$ReplicationTaskArn in the request.

$sel:assessmentResultsFile:ReplicationTaskAssessmentResult', replicationTaskAssessmentResult_assessmentResultsFile - The file containing the results of the task assessment.

$sel:replicationTaskIdentifier:ReplicationTaskAssessmentResult', replicationTaskAssessmentResult_replicationTaskIdentifier - The replication task identifier of the task on which the task assessment was run.

$sel:assessmentStatus:ReplicationTaskAssessmentResult', replicationTaskAssessmentResult_assessmentStatus - The status of the task assessment.

$sel:s3ObjectUrl:ReplicationTaskAssessmentResult', replicationTaskAssessmentResult_s3ObjectUrl - The URL of the S3 object containing the task assessment results.

The response object only contains this field if you provide DescribeReplicationTaskAssessmentResultsMessage$ReplicationTaskArn in the request.

$sel:replicationTaskLastAssessmentDate:ReplicationTaskAssessmentResult', replicationTaskAssessmentResult_replicationTaskLastAssessmentDate - The date the task assessment was completed.

$sel:replicationTaskArn:ReplicationTaskAssessmentResult', replicationTaskAssessmentResult_replicationTaskArn - The Amazon Resource Name (ARN) of the replication task.

replicationTaskAssessmentResult_assessmentResults :: Lens' ReplicationTaskAssessmentResult (Maybe Text)

The task assessment results in JSON format.

The response object only contains this field if you provide DescribeReplicationTaskAssessmentResultsMessage$ReplicationTaskArn in the request.

replicationTaskAssessmentResult_assessmentResultsFile :: Lens' ReplicationTaskAssessmentResult (Maybe Text)

The file containing the results of the task assessment.

replicationTaskAssessmentResult_replicationTaskIdentifier :: Lens' ReplicationTaskAssessmentResult (Maybe Text)

The replication task identifier of the task on which the task assessment was run.

replicationTaskAssessmentResult_assessmentStatus :: Lens' ReplicationTaskAssessmentResult (Maybe Text)

The status of the task assessment.

replicationTaskAssessmentResult_s3ObjectUrl :: Lens' ReplicationTaskAssessmentResult (Maybe Text)

The URL of the S3 object containing the task assessment results.

The response object only contains this field if you provide DescribeReplicationTaskAssessmentResultsMessage$ReplicationTaskArn in the request.

replicationTaskAssessmentResult_replicationTaskLastAssessmentDate :: Lens' ReplicationTaskAssessmentResult (Maybe UTCTime)

The date the task assessment was completed.

replicationTaskAssessmentResult_replicationTaskArn :: Lens' ReplicationTaskAssessmentResult (Maybe Text)

The Amazon Resource Name (ARN) of the replication task.