Network.AWS.DMS.Types.ReplicationTaskStats

Description

Documentation

data ReplicationTaskStats

In response to a request by the DescribeReplicationTasks operation, this object provides a collection of statistics about a replication task.

See: newReplicationTaskStats smart constructor.

Constructors

ReplicationTaskStats'

Fields

stopDate :: Maybe POSIX

The date the replication task was stopped.

fullLoadProgressPercent :: Maybe Int

The percent complete for the full load migration task.

fullLoadStartDate :: Maybe POSIX

The date the replication task full load was started.

elapsedTimeMillis :: Maybe Integer

The elapsed time of the task, in milliseconds.

startDate :: Maybe POSIX

The date the replication task was started either with a fresh start or a resume. For more information, see StartReplicationTaskType.

tablesErrored :: Maybe Int

The number of errors that have occurred during this task.

fullLoadFinishDate :: Maybe POSIX

The date the replication task full load was completed.

tablesLoaded :: Maybe Int

The number of tables loaded for this task.

tablesQueued :: Maybe Int

The number of tables queued for this task.

tablesLoading :: Maybe Int

The number of tables currently loading for this task.

freshStartDate :: Maybe POSIX

The date the replication task was started either with a fresh start or a target reload.

newReplicationTaskStats :: ReplicationTaskStats

Create a value of ReplicationTaskStats with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stopDate:ReplicationTaskStats', replicationTaskStats_stopDate - The date the replication task was stopped.

$sel:fullLoadProgressPercent:ReplicationTaskStats', replicationTaskStats_fullLoadProgressPercent - The percent complete for the full load migration task.

$sel:fullLoadStartDate:ReplicationTaskStats', replicationTaskStats_fullLoadStartDate - The date the replication task full load was started.

$sel:elapsedTimeMillis:ReplicationTaskStats', replicationTaskStats_elapsedTimeMillis - The elapsed time of the task, in milliseconds.

$sel:startDate:ReplicationTaskStats', replicationTaskStats_startDate - The date the replication task was started either with a fresh start or a resume. For more information, see StartReplicationTaskType.

$sel:tablesErrored:ReplicationTaskStats', replicationTaskStats_tablesErrored - The number of errors that have occurred during this task.

$sel:fullLoadFinishDate:ReplicationTaskStats', replicationTaskStats_fullLoadFinishDate - The date the replication task full load was completed.

$sel:tablesLoaded:ReplicationTaskStats', replicationTaskStats_tablesLoaded - The number of tables loaded for this task.

$sel:tablesQueued:ReplicationTaskStats', replicationTaskStats_tablesQueued - The number of tables queued for this task.

$sel:tablesLoading:ReplicationTaskStats', replicationTaskStats_tablesLoading - The number of tables currently loading for this task.

$sel:freshStartDate:ReplicationTaskStats', replicationTaskStats_freshStartDate - The date the replication task was started either with a fresh start or a target reload.

replicationTaskStats_stopDate :: Lens' ReplicationTaskStats (Maybe UTCTime)

The date the replication task was stopped.

replicationTaskStats_fullLoadProgressPercent :: Lens' ReplicationTaskStats (Maybe Int)

The percent complete for the full load migration task.

replicationTaskStats_fullLoadStartDate :: Lens' ReplicationTaskStats (Maybe UTCTime)

The date the replication task full load was started.

replicationTaskStats_elapsedTimeMillis :: Lens' ReplicationTaskStats (Maybe Integer)

The elapsed time of the task, in milliseconds.

replicationTaskStats_startDate :: Lens' ReplicationTaskStats (Maybe UTCTime)

The date the replication task was started either with a fresh start or a resume. For more information, see StartReplicationTaskType.

replicationTaskStats_tablesErrored :: Lens' ReplicationTaskStats (Maybe Int)

The number of errors that have occurred during this task.

replicationTaskStats_fullLoadFinishDate :: Lens' ReplicationTaskStats (Maybe UTCTime)

The date the replication task full load was completed.

replicationTaskStats_tablesLoaded :: Lens' ReplicationTaskStats (Maybe Int)

The number of tables loaded for this task.

replicationTaskStats_tablesQueued :: Lens' ReplicationTaskStats (Maybe Int)

The number of tables queued for this task.

replicationTaskStats_tablesLoading :: Lens' ReplicationTaskStats (Maybe Int)

The number of tables currently loading for this task.

replicationTaskStats_freshStartDate :: Lens' ReplicationTaskStats (Maybe UTCTime)

The date the replication task was started either with a fresh start or a target reload.