Network.AWS.DMS.Types.SupportedEndpointType

Description

Documentation

data SupportedEndpointType

Provides information about types of supported endpoints in response to a request by the DescribeEndpointTypes operation. This information includes the type of endpoint, the database engine name, and whether change data capture (CDC) is supported.

See: newSupportedEndpointType smart constructor.

Constructors

SupportedEndpointType'

Fields

replicationInstanceEngineMinimumVersion :: Maybe Text

The earliest DMS engine version that supports this endpoint engine. Note that endpoint engines released with DMS versions earlier than 3.1.1 do not return a value for this parameter.

engineName :: Maybe Text

The database engine name. Valid values, depending on the EndpointType, include "mysql", "oracle", "postgres", "mariadb", "aurora", "aurora-postgresql", "redshift", "s3", "db2", "azuredb", "sybase", "dynamodb", "mongodb", "kinesis", "kafka", "elasticsearch", "documentdb", "sqlserver", and "neptune".

endpointType :: Maybe ReplicationEndpointTypeValue

The type of endpoint. Valid values are source and target.

supportsCDC :: Maybe Bool

Indicates if change data capture (CDC) is supported.

engineDisplayName :: Maybe Text

The expanded name for the engine name. For example, if the EngineName parameter is "aurora," this value would be "Amazon Aurora MySQL."

newSupportedEndpointType :: SupportedEndpointType

Create a value of SupportedEndpointType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:replicationInstanceEngineMinimumVersion:SupportedEndpointType', supportedEndpointType_replicationInstanceEngineMinimumVersion - The earliest DMS engine version that supports this endpoint engine. Note that endpoint engines released with DMS versions earlier than 3.1.1 do not return a value for this parameter.

$sel:engineName:SupportedEndpointType', supportedEndpointType_engineName - The database engine name. Valid values, depending on the EndpointType, include "mysql", "oracle", "postgres", "mariadb", "aurora", "aurora-postgresql", "redshift", "s3", "db2", "azuredb", "sybase", "dynamodb", "mongodb", "kinesis", "kafka", "elasticsearch", "documentdb", "sqlserver", and "neptune".

$sel:endpointType:SupportedEndpointType', supportedEndpointType_endpointType - The type of endpoint. Valid values are source and target.

$sel:supportsCDC:SupportedEndpointType', supportedEndpointType_supportsCDC - Indicates if change data capture (CDC) is supported.

$sel:engineDisplayName:SupportedEndpointType', supportedEndpointType_engineDisplayName - The expanded name for the engine name. For example, if the EngineName parameter is "aurora," this value would be "Amazon Aurora MySQL."

supportedEndpointType_replicationInstanceEngineMinimumVersion :: Lens' SupportedEndpointType (Maybe Text)

The earliest DMS engine version that supports this endpoint engine. Note that endpoint engines released with DMS versions earlier than 3.1.1 do not return a value for this parameter.

supportedEndpointType_engineName :: Lens' SupportedEndpointType (Maybe Text)

The database engine name. Valid values, depending on the EndpointType, include "mysql", "oracle", "postgres", "mariadb", "aurora", "aurora-postgresql", "redshift", "s3", "db2", "azuredb", "sybase", "dynamodb", "mongodb", "kinesis", "kafka", "elasticsearch", "documentdb", "sqlserver", and "neptune".

supportedEndpointType_endpointType :: Lens' SupportedEndpointType (Maybe ReplicationEndpointTypeValue)

The type of endpoint. Valid values are source and target.

supportedEndpointType_supportsCDC :: Lens' SupportedEndpointType (Maybe Bool)

Indicates if change data capture (CDC) is supported.

supportedEndpointType_engineDisplayName :: Lens' SupportedEndpointType (Maybe Text)

The expanded name for the engine name. For example, if the EngineName parameter is "aurora," this value would be "Amazon Aurora MySQL."