Network.AWS.DMS.Types.SybaseSettings

Description

Documentation

data SybaseSettings

Provides information that defines a SAP ASE endpoint.

See: newSybaseSettings smart constructor.

Constructors

SybaseSettings'

Fields

secretsManagerSecretId :: Maybe Text

The full ARN, partial ARN, or friendly name of the SecretsManagerSecret that contains the SAP SAE endpoint connection details.

serverName :: Maybe Text

Fully qualified domain name of the endpoint.

password :: Maybe (Sensitive Text)

Endpoint connection password.

port :: Maybe Int

Endpoint TCP port.

username :: Maybe Text

Endpoint connection user name.

secretsManagerAccessRoleArn :: Maybe Text

The full Amazon Resource Name (ARN) of the IAM role that specifies DMS as the trusted entity and grants the required permissions to access the value in SecretsManagerSecret. The role must allow the iam:PassRole action. SecretsManagerSecret has the value of the Amazon Web Services Secrets Manager secret that allows access to the SAP ASE endpoint.

You can specify one of two sets of values for these permissions. You can specify the values for this setting and SecretsManagerSecretId. Or you can specify clear-text values for UserName, Password, ServerName, and Port. You can't specify both. For more information on creating this SecretsManagerSecret and the SecretsManagerAccessRoleArn and SecretsManagerSecretId required to access it, see Using secrets to access Database Migration Service resources in the Database Migration Service User Guide.

databaseName :: Maybe Text

Database name for the endpoint.

newSybaseSettings :: SybaseSettings

Create a value of SybaseSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:secretsManagerSecretId:SybaseSettings', sybaseSettings_secretsManagerSecretId - The full ARN, partial ARN, or friendly name of the SecretsManagerSecret that contains the SAP SAE endpoint connection details.

$sel:serverName:SybaseSettings', sybaseSettings_serverName - Fully qualified domain name of the endpoint.

$sel:password:SybaseSettings', sybaseSettings_password - Endpoint connection password.

$sel:port:SybaseSettings', sybaseSettings_port - Endpoint TCP port.

$sel:username:SybaseSettings', sybaseSettings_username - Endpoint connection user name.

$sel:secretsManagerAccessRoleArn:SybaseSettings', sybaseSettings_secretsManagerAccessRoleArn - The full Amazon Resource Name (ARN) of the IAM role that specifies DMS as the trusted entity and grants the required permissions to access the value in SecretsManagerSecret. The role must allow the iam:PassRole action. SecretsManagerSecret has the value of the Amazon Web Services Secrets Manager secret that allows access to the SAP ASE endpoint.

You can specify one of two sets of values for these permissions. You can specify the values for this setting and SecretsManagerSecretId. Or you can specify clear-text values for UserName, Password, ServerName, and Port. You can't specify both. For more information on creating this SecretsManagerSecret and the SecretsManagerAccessRoleArn and SecretsManagerSecretId required to access it, see Using secrets to access Database Migration Service resources in the Database Migration Service User Guide.

$sel:databaseName:SybaseSettings', sybaseSettings_databaseName - Database name for the endpoint.

sybaseSettings_secretsManagerSecretId :: Lens' SybaseSettings (Maybe Text)

The full ARN, partial ARN, or friendly name of the SecretsManagerSecret that contains the SAP SAE endpoint connection details.

sybaseSettings_serverName :: Lens' SybaseSettings (Maybe Text)

Fully qualified domain name of the endpoint.

sybaseSettings_password :: Lens' SybaseSettings (Maybe Text)

Endpoint connection password.

sybaseSettings_port :: Lens' SybaseSettings (Maybe Int)

Endpoint TCP port.

sybaseSettings_username :: Lens' SybaseSettings (Maybe Text)

Endpoint connection user name.

sybaseSettings_secretsManagerAccessRoleArn :: Lens' SybaseSettings (Maybe Text)

The full Amazon Resource Name (ARN) of the IAM role that specifies DMS as the trusted entity and grants the required permissions to access the value in SecretsManagerSecret. The role must allow the iam:PassRole action. SecretsManagerSecret has the value of the Amazon Web Services Secrets Manager secret that allows access to the SAP ASE endpoint.

You can specify one of two sets of values for these permissions. You can specify the values for this setting and SecretsManagerSecretId. Or you can specify clear-text values for UserName, Password, ServerName, and Port. You can't specify both. For more information on creating this SecretsManagerSecret and the SecretsManagerAccessRoleArn and SecretsManagerSecretId required to access it, see Using secrets to access Database Migration Service resources in the Database Migration Service User Guide.

sybaseSettings_databaseName :: Lens' SybaseSettings (Maybe Text)

Database name for the endpoint.