Network.AWS.Data.Body

Description

Documentation

getFileSize :: MonadIO m => FilePath -> m Integer

Convenience function for obtaining the size of a file.

newtype ResponseBody

A streaming, exception safe response body.

Constructors

ResponseBody

Fields

_streamBody :: ConduitM () ByteString (ResourceT IO) ()

fuseStream :: ResponseBody -> ConduitM ByteString ByteString (ResourceT IO) () -> ResponseBody

sinkBody :: MonadIO m => ResponseBody -> ConduitM ByteString Void (ResourceT IO) a -> m a

Connect a Sink to a response stream.

newtype ChunkSize

Specifies the transmitted size of the 'Transfer-Encoding' chunks.

See: defaultChunk.

Constructors

ChunkSize Int

defaultChunkSize :: ChunkSize

The default chunk size of 128 KB. The minimum chunk size accepted by AWS is 8 KB, unless the entirety of the request is below this threshold.

A chunk size of 64 KB or higher is recommended for performance reasons.

data ChunkedBody

An opaque request body which will be transmitted via Transfer-Encoding: chunked.

Invariant: Only services that support chunked encoding can accept a ChunkedBody. (Currently S3.) This is enforced by the type signatures emitted by the generator.

Constructors

ChunkedBody

Fields

_chunkedSize :: ChunkSize

_chunkedLength :: Integer

_chunkedBody :: ConduitM () ByteString (ResourceT IO) ()

chunkedLength :: Lens' ChunkedBody Integer

fuseChunks :: ChunkedBody -> ConduitM ByteString ByteString (ResourceT IO) () -> ChunkedBody

fullChunks :: ChunkedBody -> Integer

remainderBytes :: ChunkedBody -> Maybe Integer

chunkedFile :: MonadIO m => ChunkSize -> FilePath -> m RequestBody

Construct a ChunkedBody from a FilePath, where the contents will be read and signed incrementally in chunks if the target service supports it.

Will intelligently revert to HashedBody if the file is smaller than the specified ChunkSize.

See: ToBody.

chunkedFileRange

Arguments

:: MonadIO m

=> ChunkSize
The idealized size of chunks that will be yielded downstream.

-> FilePath
The file path to read.

-> Integer
The byte offset at which to start reading.

-> Integer
The maximum number of bytes to read.

-> m RequestBody

Construct a ChunkedBody from a FilePath, specifying the range of bytes to read. This can be useful for constructing multiple requests from a single file, say for S3 multipart uploads.

See: chunkedFile.

unsafeChunkedBody

Arguments

:: ChunkSize
The idealized size of chunks that will be yielded downstream.

-> Integer
The size of the stream in bytes.

-> ConduitM () ByteString (ResourceT IO) ()

-> RequestBody

Unsafely construct a ChunkedBody.

This function is marked unsafe because it does nothing to enforce the chunk size. Typically for conduit IO functions, it's whatever ByteString's defaultBufferSize is, around 32 KB. If the chunk size is less than 8 KB, the request will error. 64 KB or higher chunk size is recommended for performance reasons.

Note that it will always create a chunked body even if the request is too small.

See: ToBody.

sourceFileChunks :: MonadResource m => ChunkSize -> FilePath -> ConduitM () ByteString m ()

sourceFileRangeChunks

Arguments

:: MonadResource m

=> ChunkSize
The idealized size of chunks that will be yielded downstream.

-> FilePath
The file path to read.

-> Integer
The byte offset at which to start reading.

-> Integer
The maximum number of bytes to read.

-> ConduitM () ByteString m ()

data HashedBody

An opaque request body containing a SHA256 hash.

Constructors

HashedStream (Digest SHA256) !Integer (ConduitM () ByteString (ResourceT IO) ())

HashedBytes (Digest SHA256) ByteString

sha256Base16 :: HashedBody -> ByteString

hashedFile

Arguments

:: MonadIO m

=> FilePath
The file path to read.

-> m HashedBody

Construct a HashedBody from a FilePath, calculating the SHA256 hash and file size.

Note: While this function will perform in constant space, it will enumerate the entirety of the file contents _twice_. Firstly to calculate the SHA256 and lastly to stream the contents to the socket during sending.

See: ToHashedBody.

hashedFileRange

Arguments

:: MonadIO m

=> FilePath
The file path to read.

-> Integer
The byte offset at which to start reading.

-> Integer
The maximum number of bytes to read.

-> m HashedBody

Construct a HashedBody from a FilePath, specifying the range of bytes to read. This can be useful for constructing multiple requests from a single file, say for S3 multipart uploads.

See: hashedFile, sourceFileRange.

hashedBody

Arguments

:: Digest SHA256
A SHA256 hash of the file contents.

-> Integer
The size of the stream in bytes.

-> ConduitM () ByteString (ResourceT IO) ()

-> HashedBody

Construct a HashedBody from a Source, manually specifying the SHA256 hash and file size. It's left up to the caller to calculate these correctly, otherwise AWS will return signing errors.

See: ToHashedBody.

data RequestBody

Invariant: only services that support _both_ standard and chunked signing expose RequestBody as a parameter.

Constructors

Chunked ChunkedBody

Hashed HashedBody

md5Base64 :: RequestBody -> Maybe ByteString

isStreaming :: RequestBody -> Bool

toRequestBody :: RequestBody -> RequestBody

contentLength :: RequestBody -> Integer

class ToHashedBody a where

Anything that can be safely converted to a HashedBody.

Methods

toHashed :: a -> HashedBody

Convert a value to a hashed request body.

class ToBody a where

Anything that can be converted to a streaming request Body.

Minimal complete definition

Nothing

Methods

toBody :: a -> RequestBody

Convert a value to a request body.

default toBody :: ToHashedBody a => a -> RequestBody

_Body :: ToBody a => AReview RequestBody a