Network.AWS.DataBrew.CreateRecipeJob

Description

Creates a new job to transform input data, using steps defined in an existing Glue DataBrew recipe

Creating a Request

data CreateRecipeJob

See: newCreateRecipeJob smart constructor.

Constructors

CreateRecipeJob'

Fields

dataCatalogOutputs :: Maybe (NonEmpty DataCatalogOutput)

One or more artifacts that represent the Glue Data Catalog output from running the job.

recipeReference :: Maybe RecipeReference

databaseOutputs :: Maybe (NonEmpty DatabaseOutput)

Represents a list of JDBC database output objects which defines the output destination for a DataBrew recipe job to write to.

encryptionMode :: Maybe EncryptionMode

The encryption mode for the job, which can be one of the following:

SSE-KMS - Server-side encryption with keys managed by KMS.

SSE-S3 - Server-side encryption with keys managed by Amazon S3.

outputs :: Maybe (NonEmpty Output)

One or more artifacts that represent the output from running the job.

logSubscription :: Maybe LogSubscription

Enables or disables Amazon CloudWatch logging for the job. If logging is enabled, CloudWatch writes one log stream for each job run.

projectName :: Maybe Text

Either the name of an existing project, or a combination of a recipe and a dataset to associate with the recipe.

maxRetries :: Maybe Natural

The maximum number of times to retry the job after a job run fails.

datasetName :: Maybe Text

The name of the dataset that this job processes.

encryptionKeyArn :: Maybe Text

The Amazon Resource Name (ARN) of an encryption key that is used to protect the job.

maxCapacity :: Maybe Int

The maximum number of nodes that DataBrew can consume when the job processes data.

timeout :: Maybe Natural

The job's timeout in minutes. A job that attempts to run longer than this timeout period ends with a status of TIMEOUT.

tags :: Maybe (HashMap Text Text)

Metadata tags to apply to this job.

name :: Text

A unique name for the job. Valid characters are alphanumeric (A-Z, a-z, 0-9), hyphen (-), period (.), and space.

roleArn :: Text

The Amazon Resource Name (ARN) of the Identity and Access Management (IAM) role to be assumed when DataBrew runs the job.

newCreateRecipeJob

Arguments

:: Text
$sel:name:CreateRecipeJob'

-> Text
$sel:roleArn:CreateRecipeJob'

-> CreateRecipeJob

Create a value of CreateRecipeJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataCatalogOutputs:CreateRecipeJob', createRecipeJob_dataCatalogOutputs - One or more artifacts that represent the Glue Data Catalog output from running the job.

$sel:recipeReference:CreateRecipeJob', createRecipeJob_recipeReference - Undocumented member.

$sel:databaseOutputs:CreateRecipeJob', createRecipeJob_databaseOutputs - Represents a list of JDBC database output objects which defines the output destination for a DataBrew recipe job to write to.

$sel:encryptionMode:CreateRecipeJob', createRecipeJob_encryptionMode - The encryption mode for the job, which can be one of the following:

SSE-KMS - Server-side encryption with keys managed by KMS.

SSE-S3 - Server-side encryption with keys managed by Amazon S3.

$sel:outputs:CreateRecipeJob', createRecipeJob_outputs - One or more artifacts that represent the output from running the job.

$sel:logSubscription:CreateRecipeJob', createRecipeJob_logSubscription - Enables or disables Amazon CloudWatch logging for the job. If logging is enabled, CloudWatch writes one log stream for each job run.

$sel:projectName:CreateRecipeJob', createRecipeJob_projectName - Either the name of an existing project, or a combination of a recipe and a dataset to associate with the recipe.

$sel:maxRetries:CreateRecipeJob', createRecipeJob_maxRetries - The maximum number of times to retry the job after a job run fails.

$sel:datasetName:CreateRecipeJob', createRecipeJob_datasetName - The name of the dataset that this job processes.

$sel:encryptionKeyArn:CreateRecipeJob', createRecipeJob_encryptionKeyArn - The Amazon Resource Name (ARN) of an encryption key that is used to protect the job.

$sel:maxCapacity:CreateRecipeJob', createRecipeJob_maxCapacity - The maximum number of nodes that DataBrew can consume when the job processes data.

$sel:timeout:CreateRecipeJob', createRecipeJob_timeout - The job's timeout in minutes. A job that attempts to run longer than this timeout period ends with a status of TIMEOUT.

$sel:tags:CreateRecipeJob', createRecipeJob_tags - Metadata tags to apply to this job.

$sel:name:CreateRecipeJob', createRecipeJob_name - A unique name for the job. Valid characters are alphanumeric (A-Z, a-z, 0-9), hyphen (-), period (.), and space.

$sel:roleArn:CreateRecipeJob', createRecipeJob_roleArn - The Amazon Resource Name (ARN) of the Identity and Access Management (IAM) role to be assumed when DataBrew runs the job.

Request Lenses

createRecipeJob_dataCatalogOutputs :: Lens' CreateRecipeJob (Maybe (NonEmpty DataCatalogOutput))

One or more artifacts that represent the Glue Data Catalog output from running the job.

createRecipeJob_recipeReference :: Lens' CreateRecipeJob (Maybe RecipeReference)

Undocumented member.

createRecipeJob_databaseOutputs :: Lens' CreateRecipeJob (Maybe (NonEmpty DatabaseOutput))

Represents a list of JDBC database output objects which defines the output destination for a DataBrew recipe job to write to.

createRecipeJob_encryptionMode :: Lens' CreateRecipeJob (Maybe EncryptionMode)

The encryption mode for the job, which can be one of the following:

SSE-KMS - Server-side encryption with keys managed by KMS.

SSE-S3 - Server-side encryption with keys managed by Amazon S3.

createRecipeJob_outputs :: Lens' CreateRecipeJob (Maybe (NonEmpty Output))

One or more artifacts that represent the output from running the job.

createRecipeJob_logSubscription :: Lens' CreateRecipeJob (Maybe LogSubscription)

Enables or disables Amazon CloudWatch logging for the job. If logging is enabled, CloudWatch writes one log stream for each job run.

createRecipeJob_projectName :: Lens' CreateRecipeJob (Maybe Text)

Either the name of an existing project, or a combination of a recipe and a dataset to associate with the recipe.

createRecipeJob_maxRetries :: Lens' CreateRecipeJob (Maybe Natural)

The maximum number of times to retry the job after a job run fails.

createRecipeJob_datasetName :: Lens' CreateRecipeJob (Maybe Text)

The name of the dataset that this job processes.

createRecipeJob_encryptionKeyArn :: Lens' CreateRecipeJob (Maybe Text)

The Amazon Resource Name (ARN) of an encryption key that is used to protect the job.

createRecipeJob_maxCapacity :: Lens' CreateRecipeJob (Maybe Int)

The maximum number of nodes that DataBrew can consume when the job processes data.

createRecipeJob_timeout :: Lens' CreateRecipeJob (Maybe Natural)

The job's timeout in minutes. A job that attempts to run longer than this timeout period ends with a status of TIMEOUT.

createRecipeJob_tags :: Lens' CreateRecipeJob (Maybe (HashMap Text Text))

Metadata tags to apply to this job.

createRecipeJob_name :: Lens' CreateRecipeJob Text

A unique name for the job. Valid characters are alphanumeric (A-Z, a-z, 0-9), hyphen (-), period (.), and space.

createRecipeJob_roleArn :: Lens' CreateRecipeJob Text

The Amazon Resource Name (ARN) of the Identity and Access Management (IAM) role to be assumed when DataBrew runs the job.

Destructuring the Response

data CreateRecipeJobResponse

See: newCreateRecipeJobResponse smart constructor.

Constructors

CreateRecipeJobResponse'

Fields

httpStatus :: Int

The response's http status code.

name :: Text

The name of the job that you created.

newCreateRecipeJobResponse

Arguments

:: Int
$sel:httpStatus:CreateRecipeJobResponse'

-> Text
$sel:name:CreateRecipeJob'

-> CreateRecipeJobResponse

Create a value of CreateRecipeJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateRecipeJobResponse', createRecipeJobResponse_httpStatus - The response's http status code.

$sel:name:CreateRecipeJob', createRecipeJobResponse_name - The name of the job that you created.

Response Lenses

createRecipeJobResponse_httpStatus :: Lens' CreateRecipeJobResponse Int

The response's http status code.

createRecipeJobResponse_name :: Lens' CreateRecipeJobResponse Text

The name of the job that you created.