Network.AWS.DataBrew.CreateSchedule

Description

Creates a new schedule for one or more DataBrew jobs. Jobs can be run at a specific date and time, or at regular intervals.

Creating a Request

data CreateSchedule

See: newCreateSchedule smart constructor.

Constructors

CreateSchedule'

Fields

jobNames :: Maybe [Text]

The name or names of one or more jobs to be run.

tags :: Maybe (HashMap Text Text)

Metadata tags to apply to this schedule.

cronExpression :: Text

The date or dates and time or times when the jobs are to be run. For more information, see Cron expressions in the Glue DataBrew Developer Guide.

name :: Text

A unique name for the schedule. Valid characters are alphanumeric (A-Z, a-z, 0-9), hyphen (-), period (.), and space.

newCreateSchedule

Arguments

:: Text
$sel:cronExpression:CreateSchedule'

-> Text
$sel:name:CreateSchedule'

-> CreateSchedule

Create a value of CreateSchedule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobNames:CreateSchedule', createSchedule_jobNames - The name or names of one or more jobs to be run.

$sel:tags:CreateSchedule', createSchedule_tags - Metadata tags to apply to this schedule.

$sel:cronExpression:CreateSchedule', createSchedule_cronExpression - The date or dates and time or times when the jobs are to be run. For more information, see Cron expressions in the Glue DataBrew Developer Guide.

$sel:name:CreateSchedule', createSchedule_name - A unique name for the schedule. Valid characters are alphanumeric (A-Z, a-z, 0-9), hyphen (-), period (.), and space.

Request Lenses

createSchedule_jobNames :: Lens' CreateSchedule (Maybe [Text])

The name or names of one or more jobs to be run.

createSchedule_tags :: Lens' CreateSchedule (Maybe (HashMap Text Text))

Metadata tags to apply to this schedule.

createSchedule_cronExpression :: Lens' CreateSchedule Text

The date or dates and time or times when the jobs are to be run. For more information, see Cron expressions in the Glue DataBrew Developer Guide.

createSchedule_name :: Lens' CreateSchedule Text

A unique name for the schedule. Valid characters are alphanumeric (A-Z, a-z, 0-9), hyphen (-), period (.), and space.

Destructuring the Response

data CreateScheduleResponse

See: newCreateScheduleResponse smart constructor.

Constructors

CreateScheduleResponse'

Fields

httpStatus :: Int

The response's http status code.

name :: Text

The name of the schedule that was created.

newCreateScheduleResponse

Arguments

:: Int
$sel:httpStatus:CreateScheduleResponse'

-> Text
$sel:name:CreateSchedule'

-> CreateScheduleResponse

Create a value of CreateScheduleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateScheduleResponse', createScheduleResponse_httpStatus - The response's http status code.

$sel:name:CreateSchedule', createScheduleResponse_name - The name of the schedule that was created.

Response Lenses

createScheduleResponse_httpStatus :: Lens' CreateScheduleResponse Int

The response's http status code.

createScheduleResponse_name :: Lens' CreateScheduleResponse Text

The name of the schedule that was created.