Network.AWS.DataBrew.DescribeJob

Description

Returns the definition of a specific DataBrew job.

Creating a Request

data DescribeJob

See: newDescribeJob smart constructor.

Constructors

DescribeJob'

Fields

name :: Text

The name of the job to be described.

newDescribeJob

Arguments

:: Text
$sel:name:DescribeJob'

-> DescribeJob

Create a value of DescribeJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:DescribeJob', describeJob_name - The name of the job to be described.

Request Lenses

describeJob_name :: Lens' DescribeJob Text

The name of the job to be described.

Destructuring the Response

data DescribeJobResponse

See: newDescribeJobResponse smart constructor.

Constructors

DescribeJobResponse'

Fields

dataCatalogOutputs :: Maybe (NonEmpty DataCatalogOutput)

One or more artifacts that represent the Glue Data Catalog output from running the job.

lastModifiedDate :: Maybe POSIX

The date and time that the job was last modified.

createDate :: Maybe POSIX

The date and time that the job was created.

recipeReference :: Maybe RecipeReference

profileConfiguration :: Maybe ProfileConfiguration

Configuration for profile jobs. Used to select columns, do evaluations, and override default parameters of evaluations. When configuration is null, the profile job will run with default settings.

createdBy :: Maybe Text

The identifier (user name) of the user associated with the creation of the job.

databaseOutputs :: Maybe (NonEmpty DatabaseOutput)

Represents a list of JDBC database output objects which defines the output destination for a DataBrew recipe job to write into.

encryptionMode :: Maybe EncryptionMode

The encryption mode for the job, which can be one of the following:

SSE-KMS - Server-side encryption with keys managed by KMS.

SSE-S3 - Server-side encryption with keys managed by Amazon S3.

outputs :: Maybe (NonEmpty Output)

One or more artifacts that represent the output from running the job.

resourceArn :: Maybe Text

The Amazon Resource Name (ARN) of the job.

logSubscription :: Maybe LogSubscription

Indicates whether Amazon CloudWatch logging is enabled for this job.

projectName :: Maybe Text

The DataBrew project associated with this job.

maxRetries :: Maybe Natural

The maximum number of times to retry the job after a job run fails.

datasetName :: Maybe Text

The dataset that the job acts upon.

encryptionKeyArn :: Maybe Text

The Amazon Resource Name (ARN) of an encryption key that is used to protect the job.

maxCapacity :: Maybe Int

The maximum number of compute nodes that DataBrew can consume when the job processes data.

lastModifiedBy :: Maybe Text

The identifier (user name) of the user who last modified the job.

type' :: Maybe JobType

The job type, which must be one of the following:

PROFILE - The job analyzes the dataset to determine its size, data types, data distribution, and more.

RECIPE - The job applies one or more transformations to a dataset.

timeout :: Maybe Natural

The job's timeout in minutes. A job that attempts to run longer than this timeout period ends with a status of TIMEOUT.

tags :: Maybe (HashMap Text Text)

Metadata tags associated with this job.

roleArn :: Maybe Text

The ARN of the Identity and Access Management (IAM) role to be assumed when DataBrew runs the job.

jobSample :: Maybe JobSample

Sample configuration for profile jobs only. Determines the number of rows on which the profile job will be executed.

httpStatus :: Int

The response's http status code.

name :: Text

The name of the job.

newDescribeJobResponse

Arguments

:: Int
$sel:httpStatus:DescribeJobResponse'

-> Text
$sel:name:DescribeJob'

-> DescribeJobResponse

Create a value of DescribeJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataCatalogOutputs:DescribeJobResponse', describeJobResponse_dataCatalogOutputs - One or more artifacts that represent the Glue Data Catalog output from running the job.

$sel:lastModifiedDate:DescribeJobResponse', describeJobResponse_lastModifiedDate - The date and time that the job was last modified.

$sel:createDate:DescribeJobResponse', describeJobResponse_createDate - The date and time that the job was created.

$sel:recipeReference:DescribeJobResponse', describeJobResponse_recipeReference - Undocumented member.

$sel:profileConfiguration:DescribeJobResponse', describeJobResponse_profileConfiguration - Configuration for profile jobs. Used to select columns, do evaluations, and override default parameters of evaluations. When configuration is null, the profile job will run with default settings.

$sel:createdBy:DescribeJobResponse', describeJobResponse_createdBy - The identifier (user name) of the user associated with the creation of the job.

$sel:databaseOutputs:DescribeJobResponse', describeJobResponse_databaseOutputs - Represents a list of JDBC database output objects which defines the output destination for a DataBrew recipe job to write into.

$sel:encryptionMode:DescribeJobResponse', describeJobResponse_encryptionMode - The encryption mode for the job, which can be one of the following:

SSE-KMS - Server-side encryption with keys managed by KMS.

SSE-S3 - Server-side encryption with keys managed by Amazon S3.

$sel:outputs:DescribeJobResponse', describeJobResponse_outputs - One or more artifacts that represent the output from running the job.

$sel:resourceArn:DescribeJobResponse', describeJobResponse_resourceArn - The Amazon Resource Name (ARN) of the job.

$sel:logSubscription:DescribeJobResponse', describeJobResponse_logSubscription - Indicates whether Amazon CloudWatch logging is enabled for this job.

$sel:projectName:DescribeJobResponse', describeJobResponse_projectName - The DataBrew project associated with this job.

$sel:maxRetries:DescribeJobResponse', describeJobResponse_maxRetries - The maximum number of times to retry the job after a job run fails.

$sel:datasetName:DescribeJobResponse', describeJobResponse_datasetName - The dataset that the job acts upon.

$sel:encryptionKeyArn:DescribeJobResponse', describeJobResponse_encryptionKeyArn - The Amazon Resource Name (ARN) of an encryption key that is used to protect the job.

$sel:maxCapacity:DescribeJobResponse', describeJobResponse_maxCapacity - The maximum number of compute nodes that DataBrew can consume when the job processes data.

$sel:lastModifiedBy:DescribeJobResponse', describeJobResponse_lastModifiedBy - The identifier (user name) of the user who last modified the job.

$sel:type':DescribeJobResponse', describeJobResponse_type - The job type, which must be one of the following:

PROFILE - The job analyzes the dataset to determine its size, data types, data distribution, and more.

RECIPE - The job applies one or more transformations to a dataset.

$sel:timeout:DescribeJobResponse', describeJobResponse_timeout - The job's timeout in minutes. A job that attempts to run longer than this timeout period ends with a status of TIMEOUT.

$sel:tags:DescribeJobResponse', describeJobResponse_tags - Metadata tags associated with this job.

$sel:roleArn:DescribeJobResponse', describeJobResponse_roleArn - The ARN of the Identity and Access Management (IAM) role to be assumed when DataBrew runs the job.

$sel:jobSample:DescribeJobResponse', describeJobResponse_jobSample - Sample configuration for profile jobs only. Determines the number of rows on which the profile job will be executed.

$sel:httpStatus:DescribeJobResponse', describeJobResponse_httpStatus - The response's http status code.

$sel:name:DescribeJob', describeJobResponse_name - The name of the job.

Response Lenses

describeJobResponse_dataCatalogOutputs :: Lens' DescribeJobResponse (Maybe (NonEmpty DataCatalogOutput))

One or more artifacts that represent the Glue Data Catalog output from running the job.

describeJobResponse_lastModifiedDate :: Lens' DescribeJobResponse (Maybe UTCTime)

The date and time that the job was last modified.

describeJobResponse_createDate :: Lens' DescribeJobResponse (Maybe UTCTime)

The date and time that the job was created.

describeJobResponse_recipeReference :: Lens' DescribeJobResponse (Maybe RecipeReference)

Undocumented member.

describeJobResponse_profileConfiguration :: Lens' DescribeJobResponse (Maybe ProfileConfiguration)

Configuration for profile jobs. Used to select columns, do evaluations, and override default parameters of evaluations. When configuration is null, the profile job will run with default settings.

describeJobResponse_createdBy :: Lens' DescribeJobResponse (Maybe Text)

The identifier (user name) of the user associated with the creation of the job.

describeJobResponse_databaseOutputs :: Lens' DescribeJobResponse (Maybe (NonEmpty DatabaseOutput))

Represents a list of JDBC database output objects which defines the output destination for a DataBrew recipe job to write into.

describeJobResponse_encryptionMode :: Lens' DescribeJobResponse (Maybe EncryptionMode)

The encryption mode for the job, which can be one of the following:

SSE-KMS - Server-side encryption with keys managed by KMS.

SSE-S3 - Server-side encryption with keys managed by Amazon S3.

describeJobResponse_outputs :: Lens' DescribeJobResponse (Maybe (NonEmpty Output))

One or more artifacts that represent the output from running the job.

describeJobResponse_resourceArn :: Lens' DescribeJobResponse (Maybe Text)

The Amazon Resource Name (ARN) of the job.

describeJobResponse_logSubscription :: Lens' DescribeJobResponse (Maybe LogSubscription)

Indicates whether Amazon CloudWatch logging is enabled for this job.

describeJobResponse_projectName :: Lens' DescribeJobResponse (Maybe Text)

The DataBrew project associated with this job.

describeJobResponse_maxRetries :: Lens' DescribeJobResponse (Maybe Natural)

The maximum number of times to retry the job after a job run fails.

describeJobResponse_datasetName :: Lens' DescribeJobResponse (Maybe Text)

The dataset that the job acts upon.

describeJobResponse_encryptionKeyArn :: Lens' DescribeJobResponse (Maybe Text)

The Amazon Resource Name (ARN) of an encryption key that is used to protect the job.

describeJobResponse_maxCapacity :: Lens' DescribeJobResponse (Maybe Int)

The maximum number of compute nodes that DataBrew can consume when the job processes data.

describeJobResponse_lastModifiedBy :: Lens' DescribeJobResponse (Maybe Text)

The identifier (user name) of the user who last modified the job.

describeJobResponse_type :: Lens' DescribeJobResponse (Maybe JobType)

The job type, which must be one of the following:

PROFILE - The job analyzes the dataset to determine its size, data types, data distribution, and more.

RECIPE - The job applies one or more transformations to a dataset.

describeJobResponse_timeout :: Lens' DescribeJobResponse (Maybe Natural)

The job's timeout in minutes. A job that attempts to run longer than this timeout period ends with a status of TIMEOUT.

describeJobResponse_tags :: Lens' DescribeJobResponse (Maybe (HashMap Text Text))

Metadata tags associated with this job.

describeJobResponse_roleArn :: Lens' DescribeJobResponse (Maybe Text)

The ARN of the Identity and Access Management (IAM) role to be assumed when DataBrew runs the job.

describeJobResponse_jobSample :: Lens' DescribeJobResponse (Maybe JobSample)

Sample configuration for profile jobs only. Determines the number of rows on which the profile job will be executed.

describeJobResponse_httpStatus :: Lens' DescribeJobResponse Int

The response's http status code.

describeJobResponse_name :: Lens' DescribeJobResponse Text

The name of the job.