Network.AWS.DataBrew.DescribeProject

Description

Returns the definition of a specific DataBrew project.

Creating a Request

data DescribeProject

See: newDescribeProject smart constructor.

Constructors

DescribeProject'

Fields

name :: Text

The name of the project to be described.

newDescribeProject

Arguments

:: Text
$sel:name:DescribeProject'

-> DescribeProject

Create a value of DescribeProject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:DescribeProject', describeProject_name - The name of the project to be described.

Request Lenses

describeProject_name :: Lens' DescribeProject Text

The name of the project to be described.

Destructuring the Response

data DescribeProjectResponse

See: newDescribeProjectResponse smart constructor.

Constructors

DescribeProjectResponse'

Fields

lastModifiedDate :: Maybe POSIX

The date and time that the project was last modified.

sessionStatus :: Maybe SessionStatus

Describes the current state of the session:

PROVISIONING - allocating resources for the session.

INITIALIZING - getting the session ready for first use.

ASSIGNED - the session is ready for use.

openDate :: Maybe POSIX

The date and time when the project was opened.

createDate :: Maybe POSIX

The date and time that the project was created.

createdBy :: Maybe Text

The identifier (user name) of the user who created the project.

resourceArn :: Maybe Text

The Amazon Resource Name (ARN) of the project.

recipeName :: Maybe Text

The recipe associated with this job.

datasetName :: Maybe Text

The dataset associated with the project.

lastModifiedBy :: Maybe Text

The identifier (user name) of the user who last modified the project.

sample :: Maybe Sample

openedBy :: Maybe Text

The identifier (user name) of the user that opened the project for use.

tags :: Maybe (HashMap Text Text)

Metadata tags associated with this project.

roleArn :: Maybe Text

The ARN of the Identity and Access Management (IAM) role to be assumed when DataBrew runs the job.

httpStatus :: Int

The response's http status code.

name :: Text

The name of the project.

newDescribeProjectResponse

Arguments

:: Int
$sel:httpStatus:DescribeProjectResponse'

-> Text
$sel:name:DescribeProject'

-> DescribeProjectResponse

Create a value of DescribeProjectResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedDate:DescribeProjectResponse', describeProjectResponse_lastModifiedDate - The date and time that the project was last modified.

$sel:sessionStatus:DescribeProjectResponse', describeProjectResponse_sessionStatus - Describes the current state of the session:

PROVISIONING - allocating resources for the session.

INITIALIZING - getting the session ready for first use.

ASSIGNED - the session is ready for use.

$sel:openDate:DescribeProjectResponse', describeProjectResponse_openDate - The date and time when the project was opened.

$sel:createDate:DescribeProjectResponse', describeProjectResponse_createDate - The date and time that the project was created.

$sel:createdBy:DescribeProjectResponse', describeProjectResponse_createdBy - The identifier (user name) of the user who created the project.

$sel:resourceArn:DescribeProjectResponse', describeProjectResponse_resourceArn - The Amazon Resource Name (ARN) of the project.

$sel:recipeName:DescribeProjectResponse', describeProjectResponse_recipeName - The recipe associated with this job.

$sel:datasetName:DescribeProjectResponse', describeProjectResponse_datasetName - The dataset associated with the project.

$sel:lastModifiedBy:DescribeProjectResponse', describeProjectResponse_lastModifiedBy - The identifier (user name) of the user who last modified the project.

$sel:sample:DescribeProjectResponse', describeProjectResponse_sample - Undocumented member.

$sel:openedBy:DescribeProjectResponse', describeProjectResponse_openedBy - The identifier (user name) of the user that opened the project for use.

$sel:tags:DescribeProjectResponse', describeProjectResponse_tags - Metadata tags associated with this project.

$sel:roleArn:DescribeProjectResponse', describeProjectResponse_roleArn - The ARN of the Identity and Access Management (IAM) role to be assumed when DataBrew runs the job.

$sel:httpStatus:DescribeProjectResponse', describeProjectResponse_httpStatus - The response's http status code.

$sel:name:DescribeProject', describeProjectResponse_name - The name of the project.

Response Lenses

describeProjectResponse_lastModifiedDate :: Lens' DescribeProjectResponse (Maybe UTCTime)

The date and time that the project was last modified.

describeProjectResponse_sessionStatus :: Lens' DescribeProjectResponse (Maybe SessionStatus)

Describes the current state of the session:

PROVISIONING - allocating resources for the session.

INITIALIZING - getting the session ready for first use.

ASSIGNED - the session is ready for use.

describeProjectResponse_openDate :: Lens' DescribeProjectResponse (Maybe UTCTime)

The date and time when the project was opened.

describeProjectResponse_createDate :: Lens' DescribeProjectResponse (Maybe UTCTime)

The date and time that the project was created.

describeProjectResponse_createdBy :: Lens' DescribeProjectResponse (Maybe Text)

The identifier (user name) of the user who created the project.

describeProjectResponse_resourceArn :: Lens' DescribeProjectResponse (Maybe Text)

The Amazon Resource Name (ARN) of the project.

describeProjectResponse_recipeName :: Lens' DescribeProjectResponse (Maybe Text)

The recipe associated with this job.

describeProjectResponse_datasetName :: Lens' DescribeProjectResponse (Maybe Text)

The dataset associated with the project.

describeProjectResponse_lastModifiedBy :: Lens' DescribeProjectResponse (Maybe Text)

The identifier (user name) of the user who last modified the project.

describeProjectResponse_sample :: Lens' DescribeProjectResponse (Maybe Sample)

Undocumented member.

describeProjectResponse_openedBy :: Lens' DescribeProjectResponse (Maybe Text)

The identifier (user name) of the user that opened the project for use.

describeProjectResponse_tags :: Lens' DescribeProjectResponse (Maybe (HashMap Text Text))

Metadata tags associated with this project.

describeProjectResponse_roleArn :: Lens' DescribeProjectResponse (Maybe Text)

The ARN of the Identity and Access Management (IAM) role to be assumed when DataBrew runs the job.

describeProjectResponse_httpStatus :: Lens' DescribeProjectResponse Int

The response's http status code.

describeProjectResponse_name :: Lens' DescribeProjectResponse Text

The name of the project.