Network.AWS.DataBrew.DescribeRecipe

Description

Returns the definition of a specific DataBrew recipe corresponding to a particular version.

Creating a Request

data DescribeRecipe

See: newDescribeRecipe smart constructor.

Constructors

DescribeRecipe'

Fields

recipeVersion :: Maybe Text

The recipe version identifier. If this parameter isn't specified, then the latest published version is returned.

name :: Text

The name of the recipe to be described.

newDescribeRecipe

Arguments

:: Text
$sel:name:DescribeRecipe'

-> DescribeRecipe

Create a value of DescribeRecipe with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:recipeVersion:DescribeRecipe', describeRecipe_recipeVersion - The recipe version identifier. If this parameter isn't specified, then the latest published version is returned.

$sel:name:DescribeRecipe', describeRecipe_name - The name of the recipe to be described.

Request Lenses

describeRecipe_recipeVersion :: Lens' DescribeRecipe (Maybe Text)

The recipe version identifier. If this parameter isn't specified, then the latest published version is returned.

describeRecipe_name :: Lens' DescribeRecipe Text

The name of the recipe to be described.

Destructuring the Response

data DescribeRecipeResponse

See: newDescribeRecipeResponse smart constructor.

Constructors

DescribeRecipeResponse'

Fields

lastModifiedDate :: Maybe POSIX

The date and time that the recipe was last modified.

createDate :: Maybe POSIX

The date and time that the recipe was created.

publishedBy :: Maybe Text

The identifier (user name) of the user who last published the recipe.

createdBy :: Maybe Text

The identifier (user name) of the user who created the recipe.

steps :: Maybe [RecipeStep]

One or more steps to be performed by the recipe. Each step consists of an action, and the conditions under which the action should succeed.

publishedDate :: Maybe POSIX

The date and time when the recipe was last published.

resourceArn :: Maybe Text

The ARN of the recipe.

recipeVersion :: Maybe Text

The recipe version identifier.

projectName :: Maybe Text

The name of the project associated with this recipe.

lastModifiedBy :: Maybe Text

The identifier (user name) of the user who last modified the recipe.

description :: Maybe Text

The description of the recipe.

tags :: Maybe (HashMap Text Text)

Metadata tags associated with this project.

httpStatus :: Int

The response's http status code.

name :: Text

The name of the recipe.

newDescribeRecipeResponse

Arguments

:: Int
$sel:httpStatus:DescribeRecipeResponse'

-> Text
$sel:name:DescribeRecipe'

-> DescribeRecipeResponse

Create a value of DescribeRecipeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedDate:DescribeRecipeResponse', describeRecipeResponse_lastModifiedDate - The date and time that the recipe was last modified.

$sel:createDate:DescribeRecipeResponse', describeRecipeResponse_createDate - The date and time that the recipe was created.

$sel:publishedBy:DescribeRecipeResponse', describeRecipeResponse_publishedBy - The identifier (user name) of the user who last published the recipe.

$sel:createdBy:DescribeRecipeResponse', describeRecipeResponse_createdBy - The identifier (user name) of the user who created the recipe.

$sel:steps:DescribeRecipeResponse', describeRecipeResponse_steps - One or more steps to be performed by the recipe. Each step consists of an action, and the conditions under which the action should succeed.

$sel:publishedDate:DescribeRecipeResponse', describeRecipeResponse_publishedDate - The date and time when the recipe was last published.

$sel:resourceArn:DescribeRecipeResponse', describeRecipeResponse_resourceArn - The ARN of the recipe.

$sel:recipeVersion:DescribeRecipe', describeRecipeResponse_recipeVersion - The recipe version identifier.

$sel:projectName:DescribeRecipeResponse', describeRecipeResponse_projectName - The name of the project associated with this recipe.

$sel:lastModifiedBy:DescribeRecipeResponse', describeRecipeResponse_lastModifiedBy - The identifier (user name) of the user who last modified the recipe.

$sel:description:DescribeRecipeResponse', describeRecipeResponse_description - The description of the recipe.

$sel:tags:DescribeRecipeResponse', describeRecipeResponse_tags - Metadata tags associated with this project.

$sel:httpStatus:DescribeRecipeResponse', describeRecipeResponse_httpStatus - The response's http status code.

$sel:name:DescribeRecipe', describeRecipeResponse_name - The name of the recipe.

Response Lenses

describeRecipeResponse_lastModifiedDate :: Lens' DescribeRecipeResponse (Maybe UTCTime)

The date and time that the recipe was last modified.

describeRecipeResponse_createDate :: Lens' DescribeRecipeResponse (Maybe UTCTime)

The date and time that the recipe was created.

describeRecipeResponse_publishedBy :: Lens' DescribeRecipeResponse (Maybe Text)

The identifier (user name) of the user who last published the recipe.

describeRecipeResponse_createdBy :: Lens' DescribeRecipeResponse (Maybe Text)

The identifier (user name) of the user who created the recipe.

describeRecipeResponse_steps :: Lens' DescribeRecipeResponse (Maybe [RecipeStep])

One or more steps to be performed by the recipe. Each step consists of an action, and the conditions under which the action should succeed.

describeRecipeResponse_publishedDate :: Lens' DescribeRecipeResponse (Maybe UTCTime)

The date and time when the recipe was last published.

describeRecipeResponse_resourceArn :: Lens' DescribeRecipeResponse (Maybe Text)

The ARN of the recipe.

describeRecipeResponse_recipeVersion :: Lens' DescribeRecipeResponse (Maybe Text)

The recipe version identifier.

describeRecipeResponse_projectName :: Lens' DescribeRecipeResponse (Maybe Text)

The name of the project associated with this recipe.

describeRecipeResponse_lastModifiedBy :: Lens' DescribeRecipeResponse (Maybe Text)

The identifier (user name) of the user who last modified the recipe.

describeRecipeResponse_description :: Lens' DescribeRecipeResponse (Maybe Text)

The description of the recipe.

describeRecipeResponse_tags :: Lens' DescribeRecipeResponse (Maybe (HashMap Text Text))

Metadata tags associated with this project.

describeRecipeResponse_httpStatus :: Lens' DescribeRecipeResponse Int

The response's http status code.

describeRecipeResponse_name :: Lens' DescribeRecipeResponse Text

The name of the recipe.