Network.AWS.DataBrew.DescribeSchedule

Description

Returns the definition of a specific DataBrew schedule.

Creating a Request

data DescribeSchedule

See: newDescribeSchedule smart constructor.

Constructors

DescribeSchedule'

Fields

name :: Text

The name of the schedule to be described.

newDescribeSchedule

Arguments

:: Text
$sel:name:DescribeSchedule'

-> DescribeSchedule

Create a value of DescribeSchedule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:DescribeSchedule', describeSchedule_name - The name of the schedule to be described.

Request Lenses

describeSchedule_name :: Lens' DescribeSchedule Text

The name of the schedule to be described.

Destructuring the Response

data DescribeScheduleResponse

See: newDescribeScheduleResponse smart constructor.

Constructors

DescribeScheduleResponse'

Fields

lastModifiedDate :: Maybe POSIX

The date and time that the schedule was last modified.

createDate :: Maybe POSIX

The date and time that the schedule was created.

createdBy :: Maybe Text

The identifier (user name) of the user who created the schedule.

resourceArn :: Maybe Text

The Amazon Resource Name (ARN) of the schedule.

cronExpression :: Maybe Text

The date or dates and time or times when the jobs are to be run for the schedule. For more information, see Cron expressions in the Glue DataBrew Developer Guide.

lastModifiedBy :: Maybe Text

The identifier (user name) of the user who last modified the schedule.

jobNames :: Maybe [Text]

The name or names of one or more jobs to be run by using the schedule.

tags :: Maybe (HashMap Text Text)

Metadata tags associated with this schedule.

httpStatus :: Int

The response's http status code.

name :: Text

The name of the schedule.

newDescribeScheduleResponse

Arguments

:: Int
$sel:httpStatus:DescribeScheduleResponse'

-> Text
$sel:name:DescribeSchedule'

-> DescribeScheduleResponse

Create a value of DescribeScheduleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedDate:DescribeScheduleResponse', describeScheduleResponse_lastModifiedDate - The date and time that the schedule was last modified.

$sel:createDate:DescribeScheduleResponse', describeScheduleResponse_createDate - The date and time that the schedule was created.

$sel:createdBy:DescribeScheduleResponse', describeScheduleResponse_createdBy - The identifier (user name) of the user who created the schedule.

$sel:resourceArn:DescribeScheduleResponse', describeScheduleResponse_resourceArn - The Amazon Resource Name (ARN) of the schedule.

$sel:cronExpression:DescribeScheduleResponse', describeScheduleResponse_cronExpression - The date or dates and time or times when the jobs are to be run for the schedule. For more information, see Cron expressions in the Glue DataBrew Developer Guide.

$sel:lastModifiedBy:DescribeScheduleResponse', describeScheduleResponse_lastModifiedBy - The identifier (user name) of the user who last modified the schedule.

$sel:jobNames:DescribeScheduleResponse', describeScheduleResponse_jobNames - The name or names of one or more jobs to be run by using the schedule.

$sel:tags:DescribeScheduleResponse', describeScheduleResponse_tags - Metadata tags associated with this schedule.

$sel:httpStatus:DescribeScheduleResponse', describeScheduleResponse_httpStatus - The response's http status code.

$sel:name:DescribeSchedule', describeScheduleResponse_name - The name of the schedule.

Response Lenses

describeScheduleResponse_lastModifiedDate :: Lens' DescribeScheduleResponse (Maybe UTCTime)

The date and time that the schedule was last modified.

describeScheduleResponse_createDate :: Lens' DescribeScheduleResponse (Maybe UTCTime)

The date and time that the schedule was created.

describeScheduleResponse_createdBy :: Lens' DescribeScheduleResponse (Maybe Text)

The identifier (user name) of the user who created the schedule.

describeScheduleResponse_resourceArn :: Lens' DescribeScheduleResponse (Maybe Text)

The Amazon Resource Name (ARN) of the schedule.

describeScheduleResponse_cronExpression :: Lens' DescribeScheduleResponse (Maybe Text)

The date or dates and time or times when the jobs are to be run for the schedule. For more information, see Cron expressions in the Glue DataBrew Developer Guide.

describeScheduleResponse_lastModifiedBy :: Lens' DescribeScheduleResponse (Maybe Text)

The identifier (user name) of the user who last modified the schedule.

describeScheduleResponse_jobNames :: Lens' DescribeScheduleResponse (Maybe [Text])

The name or names of one or more jobs to be run by using the schedule.

describeScheduleResponse_tags :: Lens' DescribeScheduleResponse (Maybe (HashMap Text Text))

Metadata tags associated with this schedule.

describeScheduleResponse_httpStatus :: Lens' DescribeScheduleResponse Int

The response's http status code.

describeScheduleResponse_name :: Lens' DescribeScheduleResponse Text

The name of the schedule.