Network.AWS.DataBrew.Types.DatabaseOutput

Description

Documentation

data DatabaseOutput

Represents a JDBC database output object which defines the output destination for a DataBrew recipe job to write into.

See: newDatabaseOutput smart constructor.

Constructors

DatabaseOutput'

Fields

databaseOutputMode :: Maybe DatabaseOutputMode

The output mode to write into the database. Currently supported option: NEW_TABLE.

glueConnectionName :: Text

The Glue connection that stores the connection information for the target database.

databaseOptions :: DatabaseTableOutputOptions

Represents options that specify how and where DataBrew writes the database output generated by recipe jobs.

newDatabaseOutput

Arguments

:: Text
$sel:glueConnectionName:DatabaseOutput'

-> DatabaseTableOutputOptions
$sel:databaseOptions:DatabaseOutput'

-> DatabaseOutput

Create a value of DatabaseOutput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:databaseOutputMode:DatabaseOutput', databaseOutput_databaseOutputMode - The output mode to write into the database. Currently supported option: NEW_TABLE.

$sel:glueConnectionName:DatabaseOutput', databaseOutput_glueConnectionName - The Glue connection that stores the connection information for the target database.

$sel:databaseOptions:DatabaseOutput', databaseOutput_databaseOptions - Represents options that specify how and where DataBrew writes the database output generated by recipe jobs.

databaseOutput_databaseOutputMode :: Lens' DatabaseOutput (Maybe DatabaseOutputMode)

The output mode to write into the database. Currently supported option: NEW_TABLE.

databaseOutput_glueConnectionName :: Lens' DatabaseOutput Text

The Glue connection that stores the connection information for the target database.

databaseOutput_databaseOptions :: Lens' DatabaseOutput DatabaseTableOutputOptions

Represents options that specify how and where DataBrew writes the database output generated by recipe jobs.