Network.AWS.DataBrew.Types.Dataset

Description

Documentation

data Dataset

Represents a dataset that can be processed by DataBrew.

See: newDataset smart constructor.

Constructors

Dataset'

Fields

lastModifiedDate :: Maybe POSIX

The last modification date and time of the dataset.

pathOptions :: Maybe PathOptions

A set of options that defines how DataBrew interprets an Amazon S3 path of the dataset.

createDate :: Maybe POSIX

The date and time that the dataset was created.

formatOptions :: Maybe FormatOptions

A set of options that define how DataBrew interprets the data in the dataset.

format :: Maybe InputFormat

The file format of a dataset that is created from an Amazon S3 file or folder.

createdBy :: Maybe Text

The Amazon Resource Name (ARN) of the user who created the dataset.

accountId :: Maybe Text

The ID of the Amazon Web Services account that owns the dataset.

resourceArn :: Maybe Text

The unique Amazon Resource Name (ARN) for the dataset.

source :: Maybe Source

The location of the data for the dataset, either Amazon S3 or the Glue Data Catalog.

lastModifiedBy :: Maybe Text

The Amazon Resource Name (ARN) of the user who last modified the dataset.

tags :: Maybe (HashMap Text Text)

Metadata tags that have been applied to the dataset.

name :: Text

The unique name of the dataset.

input :: Input

Information on how DataBrew can find the dataset, in either the Glue Data Catalog or Amazon S3.

newDataset

Arguments

:: Text
$sel:name:Dataset'

-> Input
$sel:input:Dataset'

-> Dataset

Create a value of Dataset with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedDate:Dataset', dataset_lastModifiedDate - The last modification date and time of the dataset.

$sel:pathOptions:Dataset', dataset_pathOptions - A set of options that defines how DataBrew interprets an Amazon S3 path of the dataset.

$sel:createDate:Dataset', dataset_createDate - The date and time that the dataset was created.

$sel:formatOptions:Dataset', dataset_formatOptions - A set of options that define how DataBrew interprets the data in the dataset.

$sel:format:Dataset', dataset_format - The file format of a dataset that is created from an Amazon S3 file or folder.

$sel:createdBy:Dataset', dataset_createdBy - The Amazon Resource Name (ARN) of the user who created the dataset.

$sel:accountId:Dataset', dataset_accountId - The ID of the Amazon Web Services account that owns the dataset.

$sel:resourceArn:Dataset', dataset_resourceArn - The unique Amazon Resource Name (ARN) for the dataset.

$sel:source:Dataset', dataset_source - The location of the data for the dataset, either Amazon S3 or the Glue Data Catalog.

$sel:lastModifiedBy:Dataset', dataset_lastModifiedBy - The Amazon Resource Name (ARN) of the user who last modified the dataset.

$sel:tags:Dataset', dataset_tags - Metadata tags that have been applied to the dataset.

$sel:name:Dataset', dataset_name - The unique name of the dataset.

$sel:input:Dataset', dataset_input - Information on how DataBrew can find the dataset, in either the Glue Data Catalog or Amazon S3.

dataset_lastModifiedDate :: Lens' Dataset (Maybe UTCTime)

The last modification date and time of the dataset.

dataset_pathOptions :: Lens' Dataset (Maybe PathOptions)

A set of options that defines how DataBrew interprets an Amazon S3 path of the dataset.

dataset_createDate :: Lens' Dataset (Maybe UTCTime)

The date and time that the dataset was created.

dataset_formatOptions :: Lens' Dataset (Maybe FormatOptions)

A set of options that define how DataBrew interprets the data in the dataset.

dataset_format :: Lens' Dataset (Maybe InputFormat)

The file format of a dataset that is created from an Amazon S3 file or folder.

dataset_createdBy :: Lens' Dataset (Maybe Text)

The Amazon Resource Name (ARN) of the user who created the dataset.

dataset_accountId :: Lens' Dataset (Maybe Text)

The ID of the Amazon Web Services account that owns the dataset.

dataset_resourceArn :: Lens' Dataset (Maybe Text)

The unique Amazon Resource Name (ARN) for the dataset.

dataset_source :: Lens' Dataset (Maybe Source)

The location of the data for the dataset, either Amazon S3 or the Glue Data Catalog.

dataset_lastModifiedBy :: Lens' Dataset (Maybe Text)

The Amazon Resource Name (ARN) of the user who last modified the dataset.

dataset_tags :: Lens' Dataset (Maybe (HashMap Text Text))

Metadata tags that have been applied to the dataset.

dataset_name :: Lens' Dataset Text

The unique name of the dataset.

dataset_input :: Lens' Dataset Input

Information on how DataBrew can find the dataset, in either the Glue Data Catalog or Amazon S3.