Network.AWS.DataBrew.Types.JobRun

Description

Documentation

data JobRun

Represents one run of a DataBrew job.

See: newJobRun smart constructor.

Constructors

JobRun'

Fields

completedOn :: Maybe POSIX

The date and time when the job completed processing.

state :: Maybe JobRunState

The current state of the job run entity itself.

dataCatalogOutputs :: Maybe (NonEmpty DataCatalogOutput)

One or more artifacts that represent the Glue Data Catalog output from running the job.

jobName :: Maybe Text

The name of the job being processed during this run.

startedOn :: Maybe POSIX

The date and time when the job run began.

recipeReference :: Maybe RecipeReference

The set of steps processed by the job.

databaseOutputs :: Maybe (NonEmpty DatabaseOutput)

Represents a list of JDBC database output objects which defines the output destination for a DataBrew recipe job to write into.

logGroupName :: Maybe Text

The name of an Amazon CloudWatch log group, where the job writes diagnostic messages when it runs.

outputs :: Maybe (NonEmpty Output)

One or more output artifacts from a job run.

runId :: Maybe Text

The unique identifier of the job run.

executionTime :: Maybe Int

The amount of time, in seconds, during which a job run consumed resources.

logSubscription :: Maybe LogSubscription

The current status of Amazon CloudWatch logging for the job run.

startedBy :: Maybe Text

The Amazon Resource Name (ARN) of the user who initiated the job run.

datasetName :: Maybe Text

The name of the dataset for the job to process.

attempt :: Maybe Int

The number of times that DataBrew has attempted to run the job.

errorMessage :: Maybe Text

A message indicating an error (if any) that was encountered when the job ran.

jobSample :: Maybe JobSample

A sample configuration for profile jobs only, which determines the number of rows on which the profile job is run. If a JobSample value isn't provided, the default is used. The default value is CUSTOM_ROWS for the mode parameter and 20,000 for the size parameter.

newJobRun :: JobRun

Create a value of JobRun with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:completedOn:JobRun', jobRun_completedOn - The date and time when the job completed processing.

$sel:state:JobRun', jobRun_state - The current state of the job run entity itself.

$sel:dataCatalogOutputs:JobRun', jobRun_dataCatalogOutputs - One or more artifacts that represent the Glue Data Catalog output from running the job.

$sel:jobName:JobRun', jobRun_jobName - The name of the job being processed during this run.

$sel:startedOn:JobRun', jobRun_startedOn - The date and time when the job run began.

$sel:recipeReference:JobRun', jobRun_recipeReference - The set of steps processed by the job.

$sel:databaseOutputs:JobRun', jobRun_databaseOutputs - Represents a list of JDBC database output objects which defines the output destination for a DataBrew recipe job to write into.

$sel:logGroupName:JobRun', jobRun_logGroupName - The name of an Amazon CloudWatch log group, where the job writes diagnostic messages when it runs.

$sel:outputs:JobRun', jobRun_outputs - One or more output artifacts from a job run.

$sel:runId:JobRun', jobRun_runId - The unique identifier of the job run.

$sel:executionTime:JobRun', jobRun_executionTime - The amount of time, in seconds, during which a job run consumed resources.

$sel:logSubscription:JobRun', jobRun_logSubscription - The current status of Amazon CloudWatch logging for the job run.

$sel:startedBy:JobRun', jobRun_startedBy - The Amazon Resource Name (ARN) of the user who initiated the job run.

$sel:datasetName:JobRun', jobRun_datasetName - The name of the dataset for the job to process.

$sel:attempt:JobRun', jobRun_attempt - The number of times that DataBrew has attempted to run the job.

$sel:errorMessage:JobRun', jobRun_errorMessage - A message indicating an error (if any) that was encountered when the job ran.

$sel:jobSample:JobRun', jobRun_jobSample - A sample configuration for profile jobs only, which determines the number of rows on which the profile job is run. If a JobSample value isn't provided, the default is used. The default value is CUSTOM_ROWS for the mode parameter and 20,000 for the size parameter.

jobRun_completedOn :: Lens' JobRun (Maybe UTCTime)

The date and time when the job completed processing.

jobRun_state :: Lens' JobRun (Maybe JobRunState)

The current state of the job run entity itself.

jobRun_dataCatalogOutputs :: Lens' JobRun (Maybe (NonEmpty DataCatalogOutput))

One or more artifacts that represent the Glue Data Catalog output from running the job.

jobRun_jobName :: Lens' JobRun (Maybe Text)

The name of the job being processed during this run.

jobRun_startedOn :: Lens' JobRun (Maybe UTCTime)

The date and time when the job run began.

jobRun_recipeReference :: Lens' JobRun (Maybe RecipeReference)

The set of steps processed by the job.

jobRun_databaseOutputs :: Lens' JobRun (Maybe (NonEmpty DatabaseOutput))

Represents a list of JDBC database output objects which defines the output destination for a DataBrew recipe job to write into.

jobRun_logGroupName :: Lens' JobRun (Maybe Text)

The name of an Amazon CloudWatch log group, where the job writes diagnostic messages when it runs.

jobRun_outputs :: Lens' JobRun (Maybe (NonEmpty Output))

One or more output artifacts from a job run.

jobRun_runId :: Lens' JobRun (Maybe Text)

The unique identifier of the job run.

jobRun_executionTime :: Lens' JobRun (Maybe Int)

The amount of time, in seconds, during which a job run consumed resources.

jobRun_logSubscription :: Lens' JobRun (Maybe LogSubscription)

The current status of Amazon CloudWatch logging for the job run.

jobRun_startedBy :: Lens' JobRun (Maybe Text)

The Amazon Resource Name (ARN) of the user who initiated the job run.

jobRun_datasetName :: Lens' JobRun (Maybe Text)

The name of the dataset for the job to process.

jobRun_attempt :: Lens' JobRun (Maybe Int)

The number of times that DataBrew has attempted to run the job.

jobRun_errorMessage :: Lens' JobRun (Maybe Text)

A message indicating an error (if any) that was encountered when the job ran.

jobRun_jobSample :: Lens' JobRun (Maybe JobSample)

A sample configuration for profile jobs only, which determines the number of rows on which the profile job is run. If a JobSample value isn't provided, the default is used. The default value is CUSTOM_ROWS for the mode parameter and 20,000 for the size parameter.