Network.AWS.DataBrew.Types.ProfileConfiguration

Description

Documentation

data ProfileConfiguration

Configuration for profile jobs. Configuration can be used to select columns, do evaluations, and override default parameters of evaluations. When configuration is undefined, the profile job will apply default settings to all supported columns.

See: newProfileConfiguration smart constructor.

Constructors

ProfileConfiguration'

Fields

datasetStatisticsConfiguration :: Maybe StatisticsConfiguration

Configuration for inter-column evaluations. Configuration can be used to select evaluations and override parameters of evaluations. When configuration is undefined, the profile job will run all supported inter-column evaluations.

columnStatisticsConfigurations :: Maybe (NonEmpty ColumnStatisticsConfiguration)

List of configurations for column evaluations. ColumnStatisticsConfigurations are used to select evaluations and override parameters of evaluations for particular columns. When ColumnStatisticsConfigurations is undefined, the profile job will profile all supported columns and run all supported evaluations.

profileColumns :: Maybe (NonEmpty ColumnSelector)

List of column selectors. ProfileColumns can be used to select columns from the dataset. When ProfileColumns is undefined, the profile job will profile all supported columns.

newProfileConfiguration :: ProfileConfiguration

Create a value of ProfileConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:datasetStatisticsConfiguration:ProfileConfiguration', profileConfiguration_datasetStatisticsConfiguration - Configuration for inter-column evaluations. Configuration can be used to select evaluations and override parameters of evaluations. When configuration is undefined, the profile job will run all supported inter-column evaluations.

$sel:columnStatisticsConfigurations:ProfileConfiguration', profileConfiguration_columnStatisticsConfigurations - List of configurations for column evaluations. ColumnStatisticsConfigurations are used to select evaluations and override parameters of evaluations for particular columns. When ColumnStatisticsConfigurations is undefined, the profile job will profile all supported columns and run all supported evaluations.

$sel:profileColumns:ProfileConfiguration', profileConfiguration_profileColumns - List of column selectors. ProfileColumns can be used to select columns from the dataset. When ProfileColumns is undefined, the profile job will profile all supported columns.

profileConfiguration_datasetStatisticsConfiguration :: Lens' ProfileConfiguration (Maybe StatisticsConfiguration)

Configuration for inter-column evaluations. Configuration can be used to select evaluations and override parameters of evaluations. When configuration is undefined, the profile job will run all supported inter-column evaluations.

profileConfiguration_columnStatisticsConfigurations :: Lens' ProfileConfiguration (Maybe (NonEmpty ColumnStatisticsConfiguration))

List of configurations for column evaluations. ColumnStatisticsConfigurations are used to select evaluations and override parameters of evaluations for particular columns. When ColumnStatisticsConfigurations is undefined, the profile job will profile all supported columns and run all supported evaluations.

profileConfiguration_profileColumns :: Lens' ProfileConfiguration (Maybe (NonEmpty ColumnSelector))

List of column selectors. ProfileColumns can be used to select columns from the dataset. When ProfileColumns is undefined, the profile job will profile all supported columns.