Network.AWS.DataBrew.Types.RecipeVersionErrorDetail

Description

Documentation

data RecipeVersionErrorDetail

Represents any errors encountered when attempting to delete multiple recipe versions.

See: newRecipeVersionErrorDetail smart constructor.

Constructors

RecipeVersionErrorDetail'

Fields

recipeVersion :: Maybe Text

The identifier for the recipe version associated with this error.

errorCode :: Maybe Text

The HTTP status code for the error.

errorMessage :: Maybe Text

The text of the error message.

newRecipeVersionErrorDetail :: RecipeVersionErrorDetail

Create a value of RecipeVersionErrorDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:recipeVersion:RecipeVersionErrorDetail', recipeVersionErrorDetail_recipeVersion - The identifier for the recipe version associated with this error.

$sel:errorCode:RecipeVersionErrorDetail', recipeVersionErrorDetail_errorCode - The HTTP status code for the error.

$sel:errorMessage:RecipeVersionErrorDetail', recipeVersionErrorDetail_errorMessage - The text of the error message.

recipeVersionErrorDetail_recipeVersion :: Lens' RecipeVersionErrorDetail (Maybe Text)

The identifier for the recipe version associated with this error.

recipeVersionErrorDetail_errorCode :: Lens' RecipeVersionErrorDetail (Maybe Text)

The HTTP status code for the error.

recipeVersionErrorDetail_errorMessage :: Lens' RecipeVersionErrorDetail (Maybe Text)

The text of the error message.