Network.AWS.DataBrew.UpdateRecipeJob

Description

Modifies the definition of an existing DataBrew recipe job.

Creating a Request

data UpdateRecipeJob

See: newUpdateRecipeJob smart constructor.

Constructors

UpdateRecipeJob'

Fields

dataCatalogOutputs :: Maybe (NonEmpty DataCatalogOutput)

One or more artifacts that represent the Glue Data Catalog output from running the job.

databaseOutputs :: Maybe (NonEmpty DatabaseOutput)

Represents a list of JDBC database output objects which defines the output destination for a DataBrew recipe job to write into.

encryptionMode :: Maybe EncryptionMode

The encryption mode for the job, which can be one of the following:

SSE-KMS - Server-side encryption with keys managed by KMS.

SSE-S3 - Server-side encryption with keys managed by Amazon S3.

outputs :: Maybe (NonEmpty Output)

One or more artifacts that represent the output from running the job.

logSubscription :: Maybe LogSubscription

Enables or disables Amazon CloudWatch logging for the job. If logging is enabled, CloudWatch writes one log stream for each job run.

maxRetries :: Maybe Natural

The maximum number of times to retry the job after a job run fails.

encryptionKeyArn :: Maybe Text

The Amazon Resource Name (ARN) of an encryption key that is used to protect the job.

maxCapacity :: Maybe Int

The maximum number of nodes that DataBrew can consume when the job processes data.

timeout :: Maybe Natural

The job's timeout in minutes. A job that attempts to run longer than this timeout period ends with a status of TIMEOUT.

name :: Text

The name of the job to update.

roleArn :: Text

The Amazon Resource Name (ARN) of the Identity and Access Management (IAM) role to be assumed when DataBrew runs the job.

newUpdateRecipeJob

Arguments

:: Text
$sel:name:UpdateRecipeJob'

-> Text
$sel:roleArn:UpdateRecipeJob'

-> UpdateRecipeJob

Create a value of UpdateRecipeJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataCatalogOutputs:UpdateRecipeJob', updateRecipeJob_dataCatalogOutputs - One or more artifacts that represent the Glue Data Catalog output from running the job.

$sel:databaseOutputs:UpdateRecipeJob', updateRecipeJob_databaseOutputs - Represents a list of JDBC database output objects which defines the output destination for a DataBrew recipe job to write into.

$sel:encryptionMode:UpdateRecipeJob', updateRecipeJob_encryptionMode - The encryption mode for the job, which can be one of the following:

SSE-KMS - Server-side encryption with keys managed by KMS.

SSE-S3 - Server-side encryption with keys managed by Amazon S3.

$sel:outputs:UpdateRecipeJob', updateRecipeJob_outputs - One or more artifacts that represent the output from running the job.

$sel:logSubscription:UpdateRecipeJob', updateRecipeJob_logSubscription - Enables or disables Amazon CloudWatch logging for the job. If logging is enabled, CloudWatch writes one log stream for each job run.

$sel:maxRetries:UpdateRecipeJob', updateRecipeJob_maxRetries - The maximum number of times to retry the job after a job run fails.

$sel:encryptionKeyArn:UpdateRecipeJob', updateRecipeJob_encryptionKeyArn - The Amazon Resource Name (ARN) of an encryption key that is used to protect the job.

$sel:maxCapacity:UpdateRecipeJob', updateRecipeJob_maxCapacity - The maximum number of nodes that DataBrew can consume when the job processes data.

$sel:timeout:UpdateRecipeJob', updateRecipeJob_timeout - The job's timeout in minutes. A job that attempts to run longer than this timeout period ends with a status of TIMEOUT.

$sel:name:UpdateRecipeJob', updateRecipeJob_name - The name of the job to update.

$sel:roleArn:UpdateRecipeJob', updateRecipeJob_roleArn - The Amazon Resource Name (ARN) of the Identity and Access Management (IAM) role to be assumed when DataBrew runs the job.

Request Lenses

updateRecipeJob_dataCatalogOutputs :: Lens' UpdateRecipeJob (Maybe (NonEmpty DataCatalogOutput))

One or more artifacts that represent the Glue Data Catalog output from running the job.

updateRecipeJob_databaseOutputs :: Lens' UpdateRecipeJob (Maybe (NonEmpty DatabaseOutput))

Represents a list of JDBC database output objects which defines the output destination for a DataBrew recipe job to write into.

updateRecipeJob_encryptionMode :: Lens' UpdateRecipeJob (Maybe EncryptionMode)

The encryption mode for the job, which can be one of the following:

SSE-KMS - Server-side encryption with keys managed by KMS.

SSE-S3 - Server-side encryption with keys managed by Amazon S3.

updateRecipeJob_outputs :: Lens' UpdateRecipeJob (Maybe (NonEmpty Output))

One or more artifacts that represent the output from running the job.

updateRecipeJob_logSubscription :: Lens' UpdateRecipeJob (Maybe LogSubscription)

Enables or disables Amazon CloudWatch logging for the job. If logging is enabled, CloudWatch writes one log stream for each job run.

updateRecipeJob_maxRetries :: Lens' UpdateRecipeJob (Maybe Natural)

The maximum number of times to retry the job after a job run fails.

updateRecipeJob_encryptionKeyArn :: Lens' UpdateRecipeJob (Maybe Text)

The Amazon Resource Name (ARN) of an encryption key that is used to protect the job.

updateRecipeJob_maxCapacity :: Lens' UpdateRecipeJob (Maybe Int)

The maximum number of nodes that DataBrew can consume when the job processes data.

updateRecipeJob_timeout :: Lens' UpdateRecipeJob (Maybe Natural)

The job's timeout in minutes. A job that attempts to run longer than this timeout period ends with a status of TIMEOUT.

updateRecipeJob_name :: Lens' UpdateRecipeJob Text

The name of the job to update.

updateRecipeJob_roleArn :: Lens' UpdateRecipeJob Text

The Amazon Resource Name (ARN) of the Identity and Access Management (IAM) role to be assumed when DataBrew runs the job.

Destructuring the Response

data UpdateRecipeJobResponse

See: newUpdateRecipeJobResponse smart constructor.

Constructors

UpdateRecipeJobResponse'

Fields

httpStatus :: Int

The response's http status code.

name :: Text

The name of the job that you updated.

newUpdateRecipeJobResponse

Arguments

:: Int
$sel:httpStatus:UpdateRecipeJobResponse'

-> Text
$sel:name:UpdateRecipeJob'

-> UpdateRecipeJobResponse

Create a value of UpdateRecipeJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateRecipeJobResponse', updateRecipeJobResponse_httpStatus - The response's http status code.

$sel:name:UpdateRecipeJob', updateRecipeJobResponse_name - The name of the job that you updated.

Response Lenses

updateRecipeJobResponse_httpStatus :: Lens' UpdateRecipeJobResponse Int

The response's http status code.

updateRecipeJobResponse_name :: Lens' UpdateRecipeJobResponse Text

The name of the job that you updated.