Network.AWS.DataExchange.CreateEventAction

Description

This operation creates an event action.

Creating a Request

data CreateEventAction

The request body for CreateEventAction.

See: newCreateEventAction smart constructor.

Constructors

CreateEventAction'

Fields

action :: Action

What occurs after a certain event.

event :: Event

What occurs to start an action.

newCreateEventAction

Arguments

:: Action
$sel:action:CreateEventAction'

-> Event
$sel:event:CreateEventAction'

-> CreateEventAction

Create a value of CreateEventAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:action:CreateEventAction', createEventAction_action - What occurs after a certain event.

$sel:event:CreateEventAction', createEventAction_event - What occurs to start an action.

Request Lenses

createEventAction_action :: Lens' CreateEventAction Action

What occurs after a certain event.

createEventAction_event :: Lens' CreateEventAction Event

What occurs to start an action.

Destructuring the Response

data CreateEventActionResponse

See: newCreateEventActionResponse smart constructor.

Constructors

CreateEventActionResponse'

Fields

event :: Maybe Event

What occurs to start an action.

arn :: Maybe Text

The ARN for the event action.

createdAt :: Maybe POSIX

The date and time that the event action was created, in ISO 8601 format.

action :: Maybe Action

What occurs after a certain event.

id :: Maybe Text

The unique identifier for the event action.

updatedAt :: Maybe POSIX

The date and time that the event action was last updated, in ISO 8601 format.

httpStatus :: Int

The response's http status code.

newCreateEventActionResponse

Arguments

:: Int
$sel:httpStatus:CreateEventActionResponse'

-> CreateEventActionResponse

Create a value of CreateEventActionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:event:CreateEventAction', createEventActionResponse_event - What occurs to start an action.

$sel:arn:CreateEventActionResponse', createEventActionResponse_arn - The ARN for the event action.

$sel:createdAt:CreateEventActionResponse', createEventActionResponse_createdAt - The date and time that the event action was created, in ISO 8601 format.

$sel:action:CreateEventAction', createEventActionResponse_action - What occurs after a certain event.

$sel:id:CreateEventActionResponse', createEventActionResponse_id - The unique identifier for the event action.

$sel:updatedAt:CreateEventActionResponse', createEventActionResponse_updatedAt - The date and time that the event action was last updated, in ISO 8601 format.

$sel:httpStatus:CreateEventActionResponse', createEventActionResponse_httpStatus - The response's http status code.

Response Lenses

createEventActionResponse_event :: Lens' CreateEventActionResponse (Maybe Event)

What occurs to start an action.

createEventActionResponse_arn :: Lens' CreateEventActionResponse (Maybe Text)

The ARN for the event action.

createEventActionResponse_createdAt :: Lens' CreateEventActionResponse (Maybe UTCTime)

The date and time that the event action was created, in ISO 8601 format.

createEventActionResponse_action :: Lens' CreateEventActionResponse (Maybe Action)

What occurs after a certain event.

createEventActionResponse_id :: Lens' CreateEventActionResponse (Maybe Text)

The unique identifier for the event action.

createEventActionResponse_updatedAt :: Lens' CreateEventActionResponse (Maybe UTCTime)

The date and time that the event action was last updated, in ISO 8601 format.

createEventActionResponse_httpStatus :: Lens' CreateEventActionResponse Int

The response's http status code.