Network.AWS.DataExchange.CreateJob

Description

This operation creates a job.

Creating a Request

data CreateJob

The request body for CreateJob.

See: newCreateJob smart constructor.

Constructors

CreateJob'

Fields

type' :: Type

The type of job to be created.

details :: RequestDetails

The details for the CreateJob request.

newCreateJob

Arguments

:: Type
$sel:type':CreateJob'

-> RequestDetails
$sel:details:CreateJob'

-> CreateJob

Create a value of CreateJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:type':CreateJob', createJob_type - The type of job to be created.

$sel:details:CreateJob', createJob_details - The details for the CreateJob request.

Request Lenses

createJob_type :: Lens' CreateJob Type

The type of job to be created.

createJob_details :: Lens' CreateJob RequestDetails

The details for the CreateJob request.

Destructuring the Response

data CreateJobResponse

See: newCreateJobResponse smart constructor.

Constructors

CreateJobResponse'

Fields

state :: Maybe State

The state of the job.

arn :: Maybe Text

The ARN for the job.

createdAt :: Maybe POSIX

The date and time that the job was created, in ISO 8601 format.

details :: Maybe ResponseDetails

Details about the job.

id :: Maybe Text

The unique identifier for the job.

type' :: Maybe Type

The job type.

updatedAt :: Maybe POSIX

The date and time that the job was last updated, in ISO 8601 format.

errors :: Maybe [JobError]

The errors associated with jobs.

httpStatus :: Int

The response's http status code.

newCreateJobResponse

Arguments

:: Int
$sel:httpStatus:CreateJobResponse'

-> CreateJobResponse

Create a value of CreateJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:CreateJobResponse', createJobResponse_state - The state of the job.

$sel:arn:CreateJobResponse', createJobResponse_arn - The ARN for the job.

$sel:createdAt:CreateJobResponse', createJobResponse_createdAt - The date and time that the job was created, in ISO 8601 format.

$sel:details:CreateJob', createJobResponse_details - Details about the job.

$sel:id:CreateJobResponse', createJobResponse_id - The unique identifier for the job.

$sel:type':CreateJob', createJobResponse_type - The job type.

$sel:updatedAt:CreateJobResponse', createJobResponse_updatedAt - The date and time that the job was last updated, in ISO 8601 format.

$sel:errors:CreateJobResponse', createJobResponse_errors - The errors associated with jobs.

$sel:httpStatus:CreateJobResponse', createJobResponse_httpStatus - The response's http status code.

Response Lenses

createJobResponse_state :: Lens' CreateJobResponse (Maybe State)

The state of the job.

createJobResponse_arn :: Lens' CreateJobResponse (Maybe Text)

The ARN for the job.

createJobResponse_createdAt :: Lens' CreateJobResponse (Maybe UTCTime)

The date and time that the job was created, in ISO 8601 format.

createJobResponse_details :: Lens' CreateJobResponse (Maybe ResponseDetails)

Details about the job.

createJobResponse_id :: Lens' CreateJobResponse (Maybe Text)

The unique identifier for the job.

createJobResponse_type :: Lens' CreateJobResponse (Maybe Type)

The job type.

createJobResponse_updatedAt :: Lens' CreateJobResponse (Maybe UTCTime)

The date and time that the job was last updated, in ISO 8601 format.

createJobResponse_errors :: Lens' CreateJobResponse (Maybe [JobError])

The errors associated with jobs.

createJobResponse_httpStatus :: Lens' CreateJobResponse Int

The response's http status code.