Network.AWS.DataExchange.GetAsset

Description

This operation returns information about an asset.

Creating a Request

data GetAsset

See: newGetAsset smart constructor.

Constructors

GetAsset'

Fields

revisionId :: Text

The unique identifier for a revision.

assetId :: Text

The unique identifier for an asset.

dataSetId :: Text

The unique identifier for a data set.

newGetAsset

Arguments

:: Text
$sel:revisionId:GetAsset'

-> Text
$sel:assetId:GetAsset'

-> Text
$sel:dataSetId:GetAsset'

-> GetAsset

Create a value of GetAsset with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:revisionId:GetAsset', getAsset_revisionId - The unique identifier for a revision.

$sel:assetId:GetAsset', getAsset_assetId - The unique identifier for an asset.

$sel:dataSetId:GetAsset', getAsset_dataSetId - The unique identifier for a data set.

Request Lenses

getAsset_revisionId :: Lens' GetAsset Text

The unique identifier for a revision.

getAsset_assetId :: Lens' GetAsset Text

The unique identifier for an asset.

getAsset_dataSetId :: Lens' GetAsset Text

The unique identifier for a data set.

Destructuring the Response

data GetAssetResponse

See: newGetAssetResponse smart constructor.

Constructors

GetAssetResponse'

Fields

arn :: Maybe Text

The ARN for the asset.

createdAt :: Maybe POSIX

The date and time that the asset was created, in ISO 8601 format.

sourceId :: Maybe Text

The asset ID of the owned asset corresponding to the entitled asset being viewed. This parameter is returned when an asset owner is viewing the entitled copy of its owned asset.

dataSetId :: Maybe Text

The unique identifier for the data set associated with this asset.

name :: Maybe Text

The name of the asset. When importing from Amazon S3, the S3 object key is used as the asset name. When exporting to Amazon S3, the asset name is used as default target S3 object key.

assetDetails :: Maybe AssetDetails

Information about the asset.

id :: Maybe Text

The unique identifier for the asset.

assetType :: Maybe AssetType

The type of asset that is added to a data set.

updatedAt :: Maybe POSIX

The date and time that the asset was last updated, in ISO 8601 format.

revisionId :: Maybe Text

The unique identifier for the revision associated with this asset.

httpStatus :: Int

The response's http status code.

newGetAssetResponse

Arguments

:: Int
$sel:httpStatus:GetAssetResponse'

-> GetAssetResponse

Create a value of GetAssetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:GetAssetResponse', getAssetResponse_arn - The ARN for the asset.

$sel:createdAt:GetAssetResponse', getAssetResponse_createdAt - The date and time that the asset was created, in ISO 8601 format.

$sel:sourceId:GetAssetResponse', getAssetResponse_sourceId - The asset ID of the owned asset corresponding to the entitled asset being viewed. This parameter is returned when an asset owner is viewing the entitled copy of its owned asset.

$sel:dataSetId:GetAsset', getAssetResponse_dataSetId - The unique identifier for the data set associated with this asset.

$sel:name:GetAssetResponse', getAssetResponse_name - The name of the asset. When importing from Amazon S3, the S3 object key is used as the asset name. When exporting to Amazon S3, the asset name is used as default target S3 object key.

$sel:assetDetails:GetAssetResponse', getAssetResponse_assetDetails - Information about the asset.

$sel:id:GetAssetResponse', getAssetResponse_id - The unique identifier for the asset.

$sel:assetType:GetAssetResponse', getAssetResponse_assetType - The type of asset that is added to a data set.

$sel:updatedAt:GetAssetResponse', getAssetResponse_updatedAt - The date and time that the asset was last updated, in ISO 8601 format.

$sel:revisionId:GetAsset', getAssetResponse_revisionId - The unique identifier for the revision associated with this asset.

$sel:httpStatus:GetAssetResponse', getAssetResponse_httpStatus - The response's http status code.

Response Lenses

getAssetResponse_arn :: Lens' GetAssetResponse (Maybe Text)

The ARN for the asset.

getAssetResponse_createdAt :: Lens' GetAssetResponse (Maybe UTCTime)

The date and time that the asset was created, in ISO 8601 format.

getAssetResponse_sourceId :: Lens' GetAssetResponse (Maybe Text)

The asset ID of the owned asset corresponding to the entitled asset being viewed. This parameter is returned when an asset owner is viewing the entitled copy of its owned asset.

getAssetResponse_dataSetId :: Lens' GetAssetResponse (Maybe Text)

The unique identifier for the data set associated with this asset.

getAssetResponse_name :: Lens' GetAssetResponse (Maybe Text)

The name of the asset. When importing from Amazon S3, the S3 object key is used as the asset name. When exporting to Amazon S3, the asset name is used as default target S3 object key.

getAssetResponse_assetDetails :: Lens' GetAssetResponse (Maybe AssetDetails)

Information about the asset.

getAssetResponse_id :: Lens' GetAssetResponse (Maybe Text)

The unique identifier for the asset.

getAssetResponse_assetType :: Lens' GetAssetResponse (Maybe AssetType)

The type of asset that is added to a data set.

getAssetResponse_updatedAt :: Lens' GetAssetResponse (Maybe UTCTime)

The date and time that the asset was last updated, in ISO 8601 format.

getAssetResponse_revisionId :: Lens' GetAssetResponse (Maybe Text)

The unique identifier for the revision associated with this asset.

getAssetResponse_httpStatus :: Lens' GetAssetResponse Int

The response's http status code.