Network.AWS.DataExchange.GetDataSet

Description

This operation returns information about a data set.

Creating a Request

data GetDataSet

See: newGetDataSet smart constructor.

Constructors

GetDataSet'

Fields

dataSetId :: Text

The unique identifier for a data set.

newGetDataSet

Arguments

:: Text
$sel:dataSetId:GetDataSet'

-> GetDataSet

Create a value of GetDataSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataSetId:GetDataSet', getDataSet_dataSetId - The unique identifier for a data set.

Request Lenses

getDataSet_dataSetId :: Lens' GetDataSet Text

The unique identifier for a data set.

Destructuring the Response

data GetDataSetResponse

See: newGetDataSetResponse smart constructor.

Constructors

GetDataSetResponse'

Fields

origin :: Maybe Origin

A property that defines the data set as OWNED by the account (for providers) or ENTITLED to the account (for subscribers).

arn :: Maybe Text

The ARN for the data set.

createdAt :: Maybe POSIX

The date and time that the data set was created, in ISO 8601 format.

sourceId :: Maybe Text

The data set ID of the owned data set corresponding to the entitled data set being viewed. This parameter is returned when a data set owner is viewing the entitled copy of its owned data set.

originDetails :: Maybe OriginDetails

If the origin of this data set is ENTITLED, includes the details for the product on AWS Marketplace.

name :: Maybe Text

The name of the data set.

id :: Maybe Text

The unique identifier for the data set.

assetType :: Maybe AssetType

The type of asset that is added to a data set.

updatedAt :: Maybe POSIX

The date and time that the data set was last updated, in ISO 8601 format.

description :: Maybe Text

The description for the data set.

tags :: Maybe (HashMap Text Text)

The tags for the data set.

httpStatus :: Int

The response's http status code.

newGetDataSetResponse

Arguments

:: Int
$sel:httpStatus:GetDataSetResponse'

-> GetDataSetResponse

Create a value of GetDataSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:origin:GetDataSetResponse', getDataSetResponse_origin - A property that defines the data set as OWNED by the account (for providers) or ENTITLED to the account (for subscribers).

$sel:arn:GetDataSetResponse', getDataSetResponse_arn - The ARN for the data set.

$sel:createdAt:GetDataSetResponse', getDataSetResponse_createdAt - The date and time that the data set was created, in ISO 8601 format.

$sel:sourceId:GetDataSetResponse', getDataSetResponse_sourceId - The data set ID of the owned data set corresponding to the entitled data set being viewed. This parameter is returned when a data set owner is viewing the entitled copy of its owned data set.

$sel:originDetails:GetDataSetResponse', getDataSetResponse_originDetails - If the origin of this data set is ENTITLED, includes the details for the product on AWS Marketplace.

$sel:name:GetDataSetResponse', getDataSetResponse_name - The name of the data set.

$sel:id:GetDataSetResponse', getDataSetResponse_id - The unique identifier for the data set.

$sel:assetType:GetDataSetResponse', getDataSetResponse_assetType - The type of asset that is added to a data set.

$sel:updatedAt:GetDataSetResponse', getDataSetResponse_updatedAt - The date and time that the data set was last updated, in ISO 8601 format.

$sel:description:GetDataSetResponse', getDataSetResponse_description - The description for the data set.

$sel:tags:GetDataSetResponse', getDataSetResponse_tags - The tags for the data set.

$sel:httpStatus:GetDataSetResponse', getDataSetResponse_httpStatus - The response's http status code.

Response Lenses

getDataSetResponse_origin :: Lens' GetDataSetResponse (Maybe Origin)

A property that defines the data set as OWNED by the account (for providers) or ENTITLED to the account (for subscribers).

getDataSetResponse_arn :: Lens' GetDataSetResponse (Maybe Text)

The ARN for the data set.

getDataSetResponse_createdAt :: Lens' GetDataSetResponse (Maybe UTCTime)

The date and time that the data set was created, in ISO 8601 format.

getDataSetResponse_sourceId :: Lens' GetDataSetResponse (Maybe Text)

The data set ID of the owned data set corresponding to the entitled data set being viewed. This parameter is returned when a data set owner is viewing the entitled copy of its owned data set.

getDataSetResponse_originDetails :: Lens' GetDataSetResponse (Maybe OriginDetails)

If the origin of this data set is ENTITLED, includes the details for the product on AWS Marketplace.

getDataSetResponse_name :: Lens' GetDataSetResponse (Maybe Text)

The name of the data set.

getDataSetResponse_id :: Lens' GetDataSetResponse (Maybe Text)

The unique identifier for the data set.

getDataSetResponse_assetType :: Lens' GetDataSetResponse (Maybe AssetType)

The type of asset that is added to a data set.

getDataSetResponse_updatedAt :: Lens' GetDataSetResponse (Maybe UTCTime)

The date and time that the data set was last updated, in ISO 8601 format.

getDataSetResponse_description :: Lens' GetDataSetResponse (Maybe Text)

The description for the data set.

getDataSetResponse_tags :: Lens' GetDataSetResponse (Maybe (HashMap Text Text))

The tags for the data set.

getDataSetResponse_httpStatus :: Lens' GetDataSetResponse Int

The response's http status code.