Network.AWS.DataExchange.GetJob

Description

This operation returns information about a job.

Creating a Request

data GetJob

See: newGetJob smart constructor.

Constructors

GetJob'

Fields

jobId :: Text

The unique identifier for a job.

newGetJob

Arguments

:: Text
$sel:jobId:GetJob'

-> GetJob

Create a value of GetJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobId:GetJob', getJob_jobId - The unique identifier for a job.

Request Lenses

getJob_jobId :: Lens' GetJob Text

The unique identifier for a job.

Destructuring the Response

data GetJobResponse

See: newGetJobResponse smart constructor.

Constructors

GetJobResponse'

Fields

state :: Maybe State

The state of the job.

arn :: Maybe Text

The ARN for the job.

createdAt :: Maybe POSIX

The date and time that the job was created, in ISO 8601 format.

details :: Maybe ResponseDetails

Details about the job.

id :: Maybe Text

The unique identifier for the job.

type' :: Maybe Type

The job type.

updatedAt :: Maybe POSIX

The date and time that the job was last updated, in ISO 8601 format.

errors :: Maybe [JobError]

The errors associated with jobs.

httpStatus :: Int

The response's http status code.

newGetJobResponse

Arguments

:: Int
$sel:httpStatus:GetJobResponse'

-> GetJobResponse

Create a value of GetJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:GetJobResponse', getJobResponse_state - The state of the job.

$sel:arn:GetJobResponse', getJobResponse_arn - The ARN for the job.

$sel:createdAt:GetJobResponse', getJobResponse_createdAt - The date and time that the job was created, in ISO 8601 format.

$sel:details:GetJobResponse', getJobResponse_details - Details about the job.

$sel:id:GetJobResponse', getJobResponse_id - The unique identifier for the job.

$sel:type':GetJobResponse', getJobResponse_type - The job type.

$sel:updatedAt:GetJobResponse', getJobResponse_updatedAt - The date and time that the job was last updated, in ISO 8601 format.

$sel:errors:GetJobResponse', getJobResponse_errors - The errors associated with jobs.

$sel:httpStatus:GetJobResponse', getJobResponse_httpStatus - The response's http status code.

Response Lenses

getJobResponse_state :: Lens' GetJobResponse (Maybe State)

The state of the job.

getJobResponse_arn :: Lens' GetJobResponse (Maybe Text)

The ARN for the job.

getJobResponse_createdAt :: Lens' GetJobResponse (Maybe UTCTime)

The date and time that the job was created, in ISO 8601 format.

getJobResponse_details :: Lens' GetJobResponse (Maybe ResponseDetails)

Details about the job.

getJobResponse_id :: Lens' GetJobResponse (Maybe Text)

The unique identifier for the job.

getJobResponse_type :: Lens' GetJobResponse (Maybe Type)

The job type.

getJobResponse_updatedAt :: Lens' GetJobResponse (Maybe UTCTime)

The date and time that the job was last updated, in ISO 8601 format.

getJobResponse_errors :: Lens' GetJobResponse (Maybe [JobError])

The errors associated with jobs.

getJobResponse_httpStatus :: Lens' GetJobResponse Int

The response's http status code.