Network.AWS.DataExchange.GetRevision

Description

This operation returns information about a revision.

Creating a Request

data GetRevision

See: newGetRevision smart constructor.

Constructors

GetRevision'

Fields

revisionId :: Text

The unique identifier for a revision.

dataSetId :: Text

The unique identifier for a data set.

newGetRevision

Arguments

:: Text
$sel:revisionId:GetRevision'

-> Text
$sel:dataSetId:GetRevision'

-> GetRevision

Create a value of GetRevision with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:revisionId:GetRevision', getRevision_revisionId - The unique identifier for a revision.

$sel:dataSetId:GetRevision', getRevision_dataSetId - The unique identifier for a data set.

Request Lenses

getRevision_revisionId :: Lens' GetRevision Text

The unique identifier for a revision.

getRevision_dataSetId :: Lens' GetRevision Text

The unique identifier for a data set.

Destructuring the Response

data GetRevisionResponse

See: newGetRevisionResponse smart constructor.

Constructors

GetRevisionResponse'

Fields

arn :: Maybe Text

The ARN for the revision.

createdAt :: Maybe POSIX

The date and time that the revision was created, in ISO 8601 format.

sourceId :: Maybe Text

The revision ID of the owned revision corresponding to the entitled revision being viewed. This parameter is returned when a revision owner is viewing the entitled copy of its owned revision.

finalized :: Maybe Bool

To publish a revision to a data set in a product, the revision must first be finalized. Finalizing a revision tells AWS Data Exchange that your changes to the assets in the revision are complete. After it's in this read-only state, you can publish the revision to your products.

Finalized revisions can be published through the AWS Data Exchange console or the AWS Marketplace Catalog API, using the StartChangeSet AWS Marketplace Catalog API action. When using the API, revisions are uniquely identified by their ARN.

dataSetId :: Maybe Text

The unique identifier for the data set associated with this revision.

id :: Maybe Text

The unique identifier for the revision.

updatedAt :: Maybe POSIX

The date and time that the revision was last updated, in ISO 8601 format.

comment :: Maybe Text

An optional comment about the revision.

tags :: Maybe (HashMap Text Text)

The tags for the revision.

httpStatus :: Int

The response's http status code.

newGetRevisionResponse

Arguments

:: Int
$sel:httpStatus:GetRevisionResponse'

-> GetRevisionResponse

Create a value of GetRevisionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:GetRevisionResponse', getRevisionResponse_arn - The ARN for the revision.

$sel:createdAt:GetRevisionResponse', getRevisionResponse_createdAt - The date and time that the revision was created, in ISO 8601 format.

$sel:sourceId:GetRevisionResponse', getRevisionResponse_sourceId - The revision ID of the owned revision corresponding to the entitled revision being viewed. This parameter is returned when a revision owner is viewing the entitled copy of its owned revision.

$sel:finalized:GetRevisionResponse', getRevisionResponse_finalized - To publish a revision to a data set in a product, the revision must first be finalized. Finalizing a revision tells AWS Data Exchange that your changes to the assets in the revision are complete. After it's in this read-only state, you can publish the revision to your products.

Finalized revisions can be published through the AWS Data Exchange console or the AWS Marketplace Catalog API, using the StartChangeSet AWS Marketplace Catalog API action. When using the API, revisions are uniquely identified by their ARN.

$sel:dataSetId:GetRevision', getRevisionResponse_dataSetId - The unique identifier for the data set associated with this revision.

$sel:id:GetRevisionResponse', getRevisionResponse_id - The unique identifier for the revision.

$sel:updatedAt:GetRevisionResponse', getRevisionResponse_updatedAt - The date and time that the revision was last updated, in ISO 8601 format.

$sel:comment:GetRevisionResponse', getRevisionResponse_comment - An optional comment about the revision.

$sel:tags:GetRevisionResponse', getRevisionResponse_tags - The tags for the revision.

$sel:httpStatus:GetRevisionResponse', getRevisionResponse_httpStatus - The response's http status code.

Response Lenses

getRevisionResponse_arn :: Lens' GetRevisionResponse (Maybe Text)

The ARN for the revision.

getRevisionResponse_createdAt :: Lens' GetRevisionResponse (Maybe UTCTime)

The date and time that the revision was created, in ISO 8601 format.

getRevisionResponse_sourceId :: Lens' GetRevisionResponse (Maybe Text)

The revision ID of the owned revision corresponding to the entitled revision being viewed. This parameter is returned when a revision owner is viewing the entitled copy of its owned revision.

getRevisionResponse_finalized :: Lens' GetRevisionResponse (Maybe Bool)

To publish a revision to a data set in a product, the revision must first be finalized. Finalizing a revision tells AWS Data Exchange that your changes to the assets in the revision are complete. After it's in this read-only state, you can publish the revision to your products.

Finalized revisions can be published through the AWS Data Exchange console or the AWS Marketplace Catalog API, using the StartChangeSet AWS Marketplace Catalog API action. When using the API, revisions are uniquely identified by their ARN.

getRevisionResponse_dataSetId :: Lens' GetRevisionResponse (Maybe Text)

The unique identifier for the data set associated with this revision.

getRevisionResponse_id :: Lens' GetRevisionResponse (Maybe Text)

The unique identifier for the revision.

getRevisionResponse_updatedAt :: Lens' GetRevisionResponse (Maybe UTCTime)

The date and time that the revision was last updated, in ISO 8601 format.

getRevisionResponse_comment :: Lens' GetRevisionResponse (Maybe Text)

An optional comment about the revision.

getRevisionResponse_tags :: Lens' GetRevisionResponse (Maybe (HashMap Text Text))

The tags for the revision.

getRevisionResponse_httpStatus :: Lens' GetRevisionResponse Int

The response's http status code.