Network.AWS.DataExchange.ListDataSetRevisions

Description

This operation lists a data set's revisions sorted by CreatedAt in descending order.

This operation returns paginated results.

Creating a Request

data ListDataSetRevisions

See: newListDataSetRevisions smart constructor.

Constructors

ListDataSetRevisions'

Fields

nextToken :: Maybe Text

The token value retrieved from a previous call to access the next page of results.

maxResults :: Maybe Natural

The maximum number of results returned by a single call.

dataSetId :: Text

The unique identifier for a data set.

newListDataSetRevisions

Arguments

:: Text
$sel:dataSetId:ListDataSetRevisions'

-> ListDataSetRevisions

Create a value of ListDataSetRevisions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDataSetRevisions', listDataSetRevisions_nextToken - The token value retrieved from a previous call to access the next page of results.

$sel:maxResults:ListDataSetRevisions', listDataSetRevisions_maxResults - The maximum number of results returned by a single call.

$sel:dataSetId:ListDataSetRevisions', listDataSetRevisions_dataSetId - The unique identifier for a data set.

Request Lenses

listDataSetRevisions_nextToken :: Lens' ListDataSetRevisions (Maybe Text)

The token value retrieved from a previous call to access the next page of results.

listDataSetRevisions_maxResults :: Lens' ListDataSetRevisions (Maybe Natural)

The maximum number of results returned by a single call.

listDataSetRevisions_dataSetId :: Lens' ListDataSetRevisions Text

The unique identifier for a data set.

Destructuring the Response

data ListDataSetRevisionsResponse

See: newListDataSetRevisionsResponse smart constructor.

Constructors

ListDataSetRevisionsResponse'

Fields

nextToken :: Maybe Text

The token value retrieved from a previous call to access the next page of results.

revisions :: Maybe [RevisionEntry]

The asset objects listed by the request.

httpStatus :: Int

The response's http status code.

newListDataSetRevisionsResponse

Arguments

:: Int
$sel:httpStatus:ListDataSetRevisionsResponse'

-> ListDataSetRevisionsResponse

Create a value of ListDataSetRevisionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDataSetRevisions', listDataSetRevisionsResponse_nextToken - The token value retrieved from a previous call to access the next page of results.

$sel:revisions:ListDataSetRevisionsResponse', listDataSetRevisionsResponse_revisions - The asset objects listed by the request.

$sel:httpStatus:ListDataSetRevisionsResponse', listDataSetRevisionsResponse_httpStatus - The response's http status code.

Response Lenses

listDataSetRevisionsResponse_nextToken :: Lens' ListDataSetRevisionsResponse (Maybe Text)

The token value retrieved from a previous call to access the next page of results.

listDataSetRevisionsResponse_revisions :: Lens' ListDataSetRevisionsResponse (Maybe [RevisionEntry])

The asset objects listed by the request.

listDataSetRevisionsResponse_httpStatus :: Lens' ListDataSetRevisionsResponse Int

The response's http status code.