Network.AWS.DataExchange.Types.DataSetEntry

Description

Documentation

data DataSetEntry

A data set is an AWS resource with one or more revisions.

See: newDataSetEntry smart constructor.

Constructors

DataSetEntry'

Fields

sourceId :: Maybe Text

The data set ID of the owned data set corresponding to the entitled data set being viewed. This parameter is returned when a data set owner is viewing the entitled copy of its owned data set.

originDetails :: Maybe OriginDetails

If the origin of this data set is ENTITLED, includes the details for the product on AWS Marketplace.

origin :: Origin

A property that defines the data set as OWNED by the account (for providers) or ENTITLED to the account (for subscribers).

assetType :: AssetType

The type of asset that is added to a data set.

description :: Text

The description for the data set.

createdAt :: POSIX

The date and time that the data set was created, in ISO 8601 format.

id :: Text

The unique identifier for the data set.

arn :: Text

The ARN for the data set.

updatedAt :: POSIX

The date and time that the data set was last updated, in ISO 8601 format.

name :: Text

The name of the data set.

newDataSetEntry

Arguments

:: Origin
$sel:origin:DataSetEntry'

-> AssetType
$sel:assetType:DataSetEntry'

-> Text
$sel:description:DataSetEntry'

-> UTCTime
$sel:createdAt:DataSetEntry'

-> Text
$sel:id:DataSetEntry'

-> Text
$sel:arn:DataSetEntry'

-> UTCTime
$sel:updatedAt:DataSetEntry'

-> Text
$sel:name:DataSetEntry'

-> DataSetEntry

Create a value of DataSetEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceId:DataSetEntry', dataSetEntry_sourceId - The data set ID of the owned data set corresponding to the entitled data set being viewed. This parameter is returned when a data set owner is viewing the entitled copy of its owned data set.

$sel:originDetails:DataSetEntry', dataSetEntry_originDetails - If the origin of this data set is ENTITLED, includes the details for the product on AWS Marketplace.

$sel:origin:DataSetEntry', dataSetEntry_origin - A property that defines the data set as OWNED by the account (for providers) or ENTITLED to the account (for subscribers).

$sel:assetType:DataSetEntry', dataSetEntry_assetType - The type of asset that is added to a data set.

$sel:description:DataSetEntry', dataSetEntry_description - The description for the data set.

$sel:createdAt:DataSetEntry', dataSetEntry_createdAt - The date and time that the data set was created, in ISO 8601 format.

$sel:id:DataSetEntry', dataSetEntry_id - The unique identifier for the data set.

$sel:arn:DataSetEntry', dataSetEntry_arn - The ARN for the data set.

$sel:updatedAt:DataSetEntry', dataSetEntry_updatedAt - The date and time that the data set was last updated, in ISO 8601 format.

$sel:name:DataSetEntry', dataSetEntry_name - The name of the data set.

dataSetEntry_sourceId :: Lens' DataSetEntry (Maybe Text)

The data set ID of the owned data set corresponding to the entitled data set being viewed. This parameter is returned when a data set owner is viewing the entitled copy of its owned data set.

dataSetEntry_originDetails :: Lens' DataSetEntry (Maybe OriginDetails)

If the origin of this data set is ENTITLED, includes the details for the product on AWS Marketplace.

dataSetEntry_origin :: Lens' DataSetEntry Origin

A property that defines the data set as OWNED by the account (for providers) or ENTITLED to the account (for subscribers).

dataSetEntry_assetType :: Lens' DataSetEntry AssetType

The type of asset that is added to a data set.

dataSetEntry_description :: Lens' DataSetEntry Text

The description for the data set.

dataSetEntry_createdAt :: Lens' DataSetEntry UTCTime

The date and time that the data set was created, in ISO 8601 format.

dataSetEntry_id :: Lens' DataSetEntry Text

The unique identifier for the data set.

dataSetEntry_arn :: Lens' DataSetEntry Text

The ARN for the data set.

dataSetEntry_updatedAt :: Lens' DataSetEntry UTCTime

The date and time that the data set was last updated, in ISO 8601 format.

dataSetEntry_name :: Lens' DataSetEntry Text

The name of the data set.