Network.AWS.DataExchange.Types.ExportAssetToSignedUrlRequestDetails

Description

Documentation

data ExportAssetToSignedUrlRequestDetails

Details of the operation to be performed by the job.

See: newExportAssetToSignedUrlRequestDetails smart constructor.

Constructors

ExportAssetToSignedUrlRequestDetails'

Fields

dataSetId :: Text

The unique identifier for the data set associated with this export job.

assetId :: Text

The unique identifier for the asset that is exported to a signed URL.

revisionId :: Text

The unique identifier for the revision associated with this export request.

newExportAssetToSignedUrlRequestDetails

Arguments

:: Text
$sel:dataSetId:ExportAssetToSignedUrlRequestDetails'

-> Text
$sel:assetId:ExportAssetToSignedUrlRequestDetails'

-> Text
$sel:revisionId:ExportAssetToSignedUrlRequestDetails'

-> ExportAssetToSignedUrlRequestDetails

Create a value of ExportAssetToSignedUrlRequestDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataSetId:ExportAssetToSignedUrlRequestDetails', exportAssetToSignedUrlRequestDetails_dataSetId - The unique identifier for the data set associated with this export job.

$sel:assetId:ExportAssetToSignedUrlRequestDetails', exportAssetToSignedUrlRequestDetails_assetId - The unique identifier for the asset that is exported to a signed URL.

$sel:revisionId:ExportAssetToSignedUrlRequestDetails', exportAssetToSignedUrlRequestDetails_revisionId - The unique identifier for the revision associated with this export request.

exportAssetToSignedUrlRequestDetails_dataSetId :: Lens' ExportAssetToSignedUrlRequestDetails Text

The unique identifier for the data set associated with this export job.

exportAssetToSignedUrlRequestDetails_assetId :: Lens' ExportAssetToSignedUrlRequestDetails Text

The unique identifier for the asset that is exported to a signed URL.

exportAssetToSignedUrlRequestDetails_revisionId :: Lens' ExportAssetToSignedUrlRequestDetails Text

The unique identifier for the revision associated with this export request.