Network.AWS.DataExchange.Types.ExportRevisionsToS3RequestDetails

Description

Documentation

data ExportRevisionsToS3RequestDetails

Details of the operation to be performed by the job.

See: newExportRevisionsToS3RequestDetails smart constructor.

Constructors

ExportRevisionsToS3RequestDetails'

Fields

encryption :: Maybe ExportServerSideEncryption

Encryption configuration for the export job.

revisionDestinations :: [RevisionDestinationEntry]

The destination for the revision.

dataSetId :: Text

The unique identifier for the data set associated with this export job.

newExportRevisionsToS3RequestDetails

Arguments

:: Text
$sel:dataSetId:ExportRevisionsToS3RequestDetails'

-> ExportRevisionsToS3RequestDetails

Create a value of ExportRevisionsToS3RequestDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:encryption:ExportRevisionsToS3RequestDetails', exportRevisionsToS3RequestDetails_encryption - Encryption configuration for the export job.

$sel:revisionDestinations:ExportRevisionsToS3RequestDetails', exportRevisionsToS3RequestDetails_revisionDestinations - The destination for the revision.

$sel:dataSetId:ExportRevisionsToS3RequestDetails', exportRevisionsToS3RequestDetails_dataSetId - The unique identifier for the data set associated with this export job.

exportRevisionsToS3RequestDetails_encryption :: Lens' ExportRevisionsToS3RequestDetails (Maybe ExportServerSideEncryption)

Encryption configuration for the export job.

exportRevisionsToS3RequestDetails_revisionDestinations :: Lens' ExportRevisionsToS3RequestDetails [RevisionDestinationEntry]

The destination for the revision.

exportRevisionsToS3RequestDetails_dataSetId :: Lens' ExportRevisionsToS3RequestDetails Text

The unique identifier for the data set associated with this export job.