Network.AWS.DataExchange.Types.ExportServerSideEncryption

Description

Documentation

data ExportServerSideEncryption

Encryption configuration of the export job. Includes the encryption type in addition to the AWS KMS key. The KMS key is only necessary if you chose the KMS encryption. type.

See: newExportServerSideEncryption smart constructor.

Constructors

ExportServerSideEncryption'

Fields

kmsKeyArn :: Maybe Text

The Amazon Resource Name (ARN) of the AWS KMS key you want to use to encrypt the Amazon S3 objects. This parameter is required if you choose aws:kms as an encryption type.

type' :: ServerSideEncryptionTypes

The type of server side encryption used for encrypting the objects in Amazon S3.

newExportServerSideEncryption

Arguments

:: ServerSideEncryptionTypes
$sel:type':ExportServerSideEncryption'

-> ExportServerSideEncryption

Create a value of ExportServerSideEncryption with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:kmsKeyArn:ExportServerSideEncryption', exportServerSideEncryption_kmsKeyArn - The Amazon Resource Name (ARN) of the AWS KMS key you want to use to encrypt the Amazon S3 objects. This parameter is required if you choose aws:kms as an encryption type.

$sel:type':ExportServerSideEncryption', exportServerSideEncryption_type - The type of server side encryption used for encrypting the objects in Amazon S3.

exportServerSideEncryption_kmsKeyArn :: Lens' ExportServerSideEncryption (Maybe Text)

The Amazon Resource Name (ARN) of the AWS KMS key you want to use to encrypt the Amazon S3 objects. This parameter is required if you choose aws:kms as an encryption type.

exportServerSideEncryption_type :: Lens' ExportServerSideEncryption ServerSideEncryptionTypes

The type of server side encryption used for encrypting the objects in Amazon S3.