Network.AWS.DataExchange.Types.ImportAssetFromSignedUrlResponseDetails

Description

Documentation

data ImportAssetFromSignedUrlResponseDetails

The details in the response for an import request, including the signed URL and other information.

See: newImportAssetFromSignedUrlResponseDetails smart constructor.

Constructors

ImportAssetFromSignedUrlResponseDetails'

Fields

signedUrl :: Maybe Text

The signed URL.

signedUrlExpiresAt :: Maybe POSIX

The time and date at which the signed URL expires, in ISO 8601 format.

md5Hash :: Maybe Text

The Base64-encoded Md5 hash for the asset, used to ensure the integrity of the file at that location.

dataSetId :: Text

The unique identifier for the data set associated with this import job.

assetName :: Text

The name for the asset associated with this import job.

revisionId :: Text

The unique identifier for the revision associated with this import response.

newImportAssetFromSignedUrlResponseDetails

Arguments

:: Text
$sel:dataSetId:ImportAssetFromSignedUrlResponseDetails'

-> Text
$sel:assetName:ImportAssetFromSignedUrlResponseDetails'

-> Text
$sel:revisionId:ImportAssetFromSignedUrlResponseDetails'

-> ImportAssetFromSignedUrlResponseDetails

Create a value of ImportAssetFromSignedUrlResponseDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:signedUrl:ImportAssetFromSignedUrlResponseDetails', importAssetFromSignedUrlResponseDetails_signedUrl - The signed URL.

$sel:signedUrlExpiresAt:ImportAssetFromSignedUrlResponseDetails', importAssetFromSignedUrlResponseDetails_signedUrlExpiresAt - The time and date at which the signed URL expires, in ISO 8601 format.

$sel:md5Hash:ImportAssetFromSignedUrlResponseDetails', importAssetFromSignedUrlResponseDetails_md5Hash - The Base64-encoded Md5 hash for the asset, used to ensure the integrity of the file at that location.

$sel:dataSetId:ImportAssetFromSignedUrlResponseDetails', importAssetFromSignedUrlResponseDetails_dataSetId - The unique identifier for the data set associated with this import job.

$sel:assetName:ImportAssetFromSignedUrlResponseDetails', importAssetFromSignedUrlResponseDetails_assetName - The name for the asset associated with this import job.

$sel:revisionId:ImportAssetFromSignedUrlResponseDetails', importAssetFromSignedUrlResponseDetails_revisionId - The unique identifier for the revision associated with this import response.

importAssetFromSignedUrlResponseDetails_signedUrl :: Lens' ImportAssetFromSignedUrlResponseDetails (Maybe Text)

The signed URL.

importAssetFromSignedUrlResponseDetails_signedUrlExpiresAt :: Lens' ImportAssetFromSignedUrlResponseDetails (Maybe UTCTime)

The time and date at which the signed URL expires, in ISO 8601 format.

importAssetFromSignedUrlResponseDetails_md5Hash :: Lens' ImportAssetFromSignedUrlResponseDetails (Maybe Text)

The Base64-encoded Md5 hash for the asset, used to ensure the integrity of the file at that location.

importAssetFromSignedUrlResponseDetails_dataSetId :: Lens' ImportAssetFromSignedUrlResponseDetails Text

The unique identifier for the data set associated with this import job.

importAssetFromSignedUrlResponseDetails_assetName :: Lens' ImportAssetFromSignedUrlResponseDetails Text

The name for the asset associated with this import job.

importAssetFromSignedUrlResponseDetails_revisionId :: Lens' ImportAssetFromSignedUrlResponseDetails Text

The unique identifier for the revision associated with this import response.