Network.AWS.DataExchange.Types.ResponseDetails

Description

Documentation

data ResponseDetails

Details for the response.

See: newResponseDetails smart constructor.

Constructors

ResponseDetails'

Fields

exportAssetsToS3 :: Maybe ExportAssetsToS3ResponseDetails

Details for the export to Amazon S3 response.

exportRevisionsToS3 :: Maybe ExportRevisionsToS3ResponseDetails

Details for the export revisions to Amazon S3 response.

importAssetFromSignedUrl :: Maybe ImportAssetFromSignedUrlResponseDetails

Details for the import from signed URL response.

importAssetsFromRedshiftDataShares :: Maybe ImportAssetsFromRedshiftDataSharesResponseDetails

Details from an import from Amazon Redshift datashare response.

importAssetsFromS3 :: Maybe ImportAssetsFromS3ResponseDetails

Details for the import from Amazon S3 response.

exportAssetToSignedUrl :: Maybe ExportAssetToSignedUrlResponseDetails

Details for the export to signed URL response.

newResponseDetails :: ResponseDetails

Create a value of ResponseDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:exportAssetsToS3:ResponseDetails', responseDetails_exportAssetsToS3 - Details for the export to Amazon S3 response.

$sel:exportRevisionsToS3:ResponseDetails', responseDetails_exportRevisionsToS3 - Details for the export revisions to Amazon S3 response.

$sel:importAssetFromSignedUrl:ResponseDetails', responseDetails_importAssetFromSignedUrl - Details for the import from signed URL response.

$sel:importAssetsFromRedshiftDataShares:ResponseDetails', responseDetails_importAssetsFromRedshiftDataShares - Details from an import from Amazon Redshift datashare response.

$sel:importAssetsFromS3:ResponseDetails', responseDetails_importAssetsFromS3 - Details for the import from Amazon S3 response.

$sel:exportAssetToSignedUrl:ResponseDetails', responseDetails_exportAssetToSignedUrl - Details for the export to signed URL response.

responseDetails_exportAssetsToS3 :: Lens' ResponseDetails (Maybe ExportAssetsToS3ResponseDetails)

Details for the export to Amazon S3 response.

responseDetails_exportRevisionsToS3 :: Lens' ResponseDetails (Maybe ExportRevisionsToS3ResponseDetails)

Details for the export revisions to Amazon S3 response.

responseDetails_importAssetFromSignedUrl :: Lens' ResponseDetails (Maybe ImportAssetFromSignedUrlResponseDetails)

Details for the import from signed URL response.

responseDetails_importAssetsFromRedshiftDataShares :: Lens' ResponseDetails (Maybe ImportAssetsFromRedshiftDataSharesResponseDetails)

Details from an import from Amazon Redshift datashare response.

responseDetails_importAssetsFromS3 :: Lens' ResponseDetails (Maybe ImportAssetsFromS3ResponseDetails)

Details for the import from Amazon S3 response.

responseDetails_exportAssetToSignedUrl :: Lens' ResponseDetails (Maybe ExportAssetToSignedUrlResponseDetails)

Details for the export to signed URL response.