Network.AWS.DataExchange.Types.RevisionEntry

Description

Documentation

data RevisionEntry

A revision is a container for one or more assets.

See: newRevisionEntry smart constructor.

Constructors

RevisionEntry'

Fields

sourceId :: Maybe Text

The revision ID of the owned revision corresponding to the entitled revision being viewed. This parameter is returned when a revision owner is viewing the entitled copy of its owned revision.

finalized :: Maybe Bool

To publish a revision to a data set in a product, the revision must first be finalized. Finalizing a revision tells AWS Data Exchange that your changes to the assets in the revision are complete. After it's in this read-only state, you can publish the revision to your products.

Finalized revisions can be published through the AWS Data Exchange console or the AWS Marketplace Catalog API, using the StartChangeSet AWS Marketplace Catalog API action. When using the API, revisions are uniquely identified by their ARN.

comment :: Maybe Text

An optional comment about the revision.

createdAt :: POSIX

The date and time that the revision was created, in ISO 8601 format.

dataSetId :: Text

The unique identifier for the data set associated with this revision.

id :: Text

The unique identifier for the revision.

arn :: Text

The ARN for the revision.

updatedAt :: POSIX

The date and time that the revision was last updated, in ISO 8601 format.

newRevisionEntry

Arguments

:: UTCTime
$sel:createdAt:RevisionEntry'

-> Text
$sel:dataSetId:RevisionEntry'

-> Text
$sel:id:RevisionEntry'

-> Text
$sel:arn:RevisionEntry'

-> UTCTime
$sel:updatedAt:RevisionEntry'

-> RevisionEntry

Create a value of RevisionEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceId:RevisionEntry', revisionEntry_sourceId - The revision ID of the owned revision corresponding to the entitled revision being viewed. This parameter is returned when a revision owner is viewing the entitled copy of its owned revision.

$sel:finalized:RevisionEntry', revisionEntry_finalized - To publish a revision to a data set in a product, the revision must first be finalized. Finalizing a revision tells AWS Data Exchange that your changes to the assets in the revision are complete. After it's in this read-only state, you can publish the revision to your products.

Finalized revisions can be published through the AWS Data Exchange console or the AWS Marketplace Catalog API, using the StartChangeSet AWS Marketplace Catalog API action. When using the API, revisions are uniquely identified by their ARN.

$sel:comment:RevisionEntry', revisionEntry_comment - An optional comment about the revision.

$sel:createdAt:RevisionEntry', revisionEntry_createdAt - The date and time that the revision was created, in ISO 8601 format.

$sel:dataSetId:RevisionEntry', revisionEntry_dataSetId - The unique identifier for the data set associated with this revision.

$sel:id:RevisionEntry', revisionEntry_id - The unique identifier for the revision.

$sel:arn:RevisionEntry', revisionEntry_arn - The ARN for the revision.

$sel:updatedAt:RevisionEntry', revisionEntry_updatedAt - The date and time that the revision was last updated, in ISO 8601 format.

revisionEntry_sourceId :: Lens' RevisionEntry (Maybe Text)

The revision ID of the owned revision corresponding to the entitled revision being viewed. This parameter is returned when a revision owner is viewing the entitled copy of its owned revision.

revisionEntry_finalized :: Lens' RevisionEntry (Maybe Bool)

To publish a revision to a data set in a product, the revision must first be finalized. Finalizing a revision tells AWS Data Exchange that your changes to the assets in the revision are complete. After it's in this read-only state, you can publish the revision to your products.

Finalized revisions can be published through the AWS Data Exchange console or the AWS Marketplace Catalog API, using the StartChangeSet AWS Marketplace Catalog API action. When using the API, revisions are uniquely identified by their ARN.

revisionEntry_comment :: Lens' RevisionEntry (Maybe Text)

An optional comment about the revision.

revisionEntry_createdAt :: Lens' RevisionEntry UTCTime

The date and time that the revision was created, in ISO 8601 format.

revisionEntry_dataSetId :: Lens' RevisionEntry Text

The unique identifier for the data set associated with this revision.

revisionEntry_id :: Lens' RevisionEntry Text

The unique identifier for the revision.

revisionEntry_arn :: Lens' RevisionEntry Text

The ARN for the revision.

revisionEntry_updatedAt :: Lens' RevisionEntry UTCTime

The date and time that the revision was last updated, in ISO 8601 format.