Network.AWS.DataExchange.UpdateDataSet

Description

This operation updates a data set.

Creating a Request

data UpdateDataSet

The request body for UpdateDataSet.

See: newUpdateDataSet smart constructor.

Constructors

UpdateDataSet'

Fields

name :: Maybe Text

The name of the data set.

description :: Maybe Text

The description for the data set.

dataSetId :: Text

The unique identifier for a data set.

newUpdateDataSet

Arguments

:: Text
$sel:dataSetId:UpdateDataSet'

-> UpdateDataSet

Create a value of UpdateDataSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:UpdateDataSet', updateDataSet_name - The name of the data set.

$sel:description:UpdateDataSet', updateDataSet_description - The description for the data set.

$sel:dataSetId:UpdateDataSet', updateDataSet_dataSetId - The unique identifier for a data set.

Request Lenses

updateDataSet_name :: Lens' UpdateDataSet (Maybe Text)

The name of the data set.

updateDataSet_description :: Lens' UpdateDataSet (Maybe Text)

The description for the data set.

updateDataSet_dataSetId :: Lens' UpdateDataSet Text

The unique identifier for a data set.

Destructuring the Response

data UpdateDataSetResponse

See: newUpdateDataSetResponse smart constructor.

Constructors

UpdateDataSetResponse'

Fields

origin :: Maybe Origin

A property that defines the data set as OWNED by the account (for providers) or ENTITLED to the account (for subscribers).

arn :: Maybe Text

The ARN for the data set.

createdAt :: Maybe POSIX

The date and time that the data set was created, in ISO 8601 format.

sourceId :: Maybe Text

The data set ID of the owned data set corresponding to the entitled data set being viewed. This parameter is returned when a data set owner is viewing the entitled copy of its owned data set.

originDetails :: Maybe OriginDetails

If the origin of this data set is ENTITLED, includes the details for the product on AWS Marketplace.

name :: Maybe Text

The name of the data set.

id :: Maybe Text

The unique identifier for the data set.

assetType :: Maybe AssetType

The type of asset that is added to a data set.

updatedAt :: Maybe POSIX

The date and time that the data set was last updated, in ISO 8601 format.

description :: Maybe Text

The description for the data set.

httpStatus :: Int

The response's http status code.

newUpdateDataSetResponse

Arguments

:: Int
$sel:httpStatus:UpdateDataSetResponse'

-> UpdateDataSetResponse

Create a value of UpdateDataSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:origin:UpdateDataSetResponse', updateDataSetResponse_origin - A property that defines the data set as OWNED by the account (for providers) or ENTITLED to the account (for subscribers).

$sel:arn:UpdateDataSetResponse', updateDataSetResponse_arn - The ARN for the data set.

$sel:createdAt:UpdateDataSetResponse', updateDataSetResponse_createdAt - The date and time that the data set was created, in ISO 8601 format.

$sel:sourceId:UpdateDataSetResponse', updateDataSetResponse_sourceId - The data set ID of the owned data set corresponding to the entitled data set being viewed. This parameter is returned when a data set owner is viewing the entitled copy of its owned data set.

$sel:originDetails:UpdateDataSetResponse', updateDataSetResponse_originDetails - If the origin of this data set is ENTITLED, includes the details for the product on AWS Marketplace.

$sel:name:UpdateDataSet', updateDataSetResponse_name - The name of the data set.

$sel:id:UpdateDataSetResponse', updateDataSetResponse_id - The unique identifier for the data set.

$sel:assetType:UpdateDataSetResponse', updateDataSetResponse_assetType - The type of asset that is added to a data set.

$sel:updatedAt:UpdateDataSetResponse', updateDataSetResponse_updatedAt - The date and time that the data set was last updated, in ISO 8601 format.

$sel:description:UpdateDataSet', updateDataSetResponse_description - The description for the data set.

$sel:httpStatus:UpdateDataSetResponse', updateDataSetResponse_httpStatus - The response's http status code.

Response Lenses

updateDataSetResponse_origin :: Lens' UpdateDataSetResponse (Maybe Origin)

A property that defines the data set as OWNED by the account (for providers) or ENTITLED to the account (for subscribers).

updateDataSetResponse_arn :: Lens' UpdateDataSetResponse (Maybe Text)

The ARN for the data set.

updateDataSetResponse_createdAt :: Lens' UpdateDataSetResponse (Maybe UTCTime)

The date and time that the data set was created, in ISO 8601 format.

updateDataSetResponse_sourceId :: Lens' UpdateDataSetResponse (Maybe Text)

The data set ID of the owned data set corresponding to the entitled data set being viewed. This parameter is returned when a data set owner is viewing the entitled copy of its owned data set.

updateDataSetResponse_originDetails :: Lens' UpdateDataSetResponse (Maybe OriginDetails)

If the origin of this data set is ENTITLED, includes the details for the product on AWS Marketplace.

updateDataSetResponse_name :: Lens' UpdateDataSetResponse (Maybe Text)

The name of the data set.

updateDataSetResponse_id :: Lens' UpdateDataSetResponse (Maybe Text)

The unique identifier for the data set.

updateDataSetResponse_assetType :: Lens' UpdateDataSetResponse (Maybe AssetType)

The type of asset that is added to a data set.

updateDataSetResponse_updatedAt :: Lens' UpdateDataSetResponse (Maybe UTCTime)

The date and time that the data set was last updated, in ISO 8601 format.

updateDataSetResponse_description :: Lens' UpdateDataSetResponse (Maybe Text)

The description for the data set.

updateDataSetResponse_httpStatus :: Lens' UpdateDataSetResponse Int

The response's http status code.