Network.AWS.DataPipeline

Description

Derived from API version 2012-10-29 of the AWS service descriptions, licensed under Apache 2.0.

AWS Data Pipeline configures and manages a data-driven workflow called a pipeline. AWS Data Pipeline handles the details of scheduling and ensuring that data dependencies are met so that your application can focus on processing the data.

AWS Data Pipeline provides a JAR implementation of a task runner called AWS Data Pipeline Task Runner. AWS Data Pipeline Task Runner provides logic for common data management scenarios, such as performing database queries and running data analysis using Amazon Elastic MapReduce (Amazon EMR). You can use AWS Data Pipeline Task Runner as your task runner, or you can write your own task runner to provide custom data management.

AWS Data Pipeline implements two main sets of functionality. Use the first set to create a pipeline and define data sources, schedules, dependencies, and the transforms to be performed on the data. Use the second set in your task runner application to receive the next task ready for processing. The logic for performing the task, such as querying the data, running data analysis, or converting the data from one format to another, is contained within the task runner. The task runner performs the task assigned to it by the web service, reporting progress to the web service as it does so. When the task is done, the task runner reports the final success or failure of the task to the web service.

Service Configuration

defaultService :: Service

API version 2012-10-29 of the Amazon Data Pipeline SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by DataPipeline.

TaskNotFoundException

_TaskNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified task was not found.

InvalidRequestException

_InvalidRequestException :: AsError a => Getting (First ServiceError) a ServiceError

The request was not valid. Verify that your request was properly formatted, that the signature was generated with the correct credentials, and that you haven't exceeded any of the service limits for your account.

PipelineNotFoundException

_PipelineNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified pipeline was not found. Verify that you used the correct user and account identifiers.

PipelineDeletedException

_PipelineDeletedException :: AsError a => Getting (First ServiceError) a ServiceError

The specified pipeline has been deleted.

InternalServiceError

_InternalServiceError :: AsError a => Getting (First ServiceError) a ServiceError

An internal service error occurred.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

RemoveTags

data RemoveTags

Contains the parameters for RemoveTags.

See: newRemoveTags smart constructor.

Constructors

RemoveTags' Text [Text]

newRemoveTags

Arguments

:: Text
$sel:pipelineId:RemoveTags'

-> RemoveTags

Create a value of RemoveTags with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pipelineId:RemoveTags', removeTags_pipelineId - The ID of the pipeline.

$sel:tagKeys:RemoveTags', removeTags_tagKeys - The keys of the tags to remove.

data RemoveTagsResponse

Contains the output of RemoveTags.

See: newRemoveTagsResponse smart constructor.

Constructors

RemoveTagsResponse' Int

newRemoveTagsResponse

Arguments

:: Int
$sel:httpStatus:RemoveTagsResponse'

-> RemoveTagsResponse

Create a value of RemoveTagsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:RemoveTagsResponse', removeTagsResponse_httpStatus - The response's http status code.

QueryObjects (Paginated)

data QueryObjects

Contains the parameters for QueryObjects.

See: newQueryObjects smart constructor.

Constructors

QueryObjects' (Maybe Query) (Maybe Int) (Maybe Text) Text Text

newQueryObjects

Arguments

:: Text
$sel:pipelineId:QueryObjects'

-> Text
$sel:sphere:QueryObjects'

-> QueryObjects

Create a value of QueryObjects with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:query:QueryObjects', queryObjects_query - The query that defines the objects to be returned. The Query object can contain a maximum of ten selectors. The conditions in the query are limited to top-level String fields in the object. These filters can be applied to components, instances, and attempts.

$sel:limit:QueryObjects', queryObjects_limit - The maximum number of object names that QueryObjects will return in a single call. The default value is 100.

$sel:marker:QueryObjects', queryObjects_marker - The starting point for the results to be returned. For the first call, this value should be empty. As long as there are more results, continue to call QueryObjects with the marker value from the previous call to retrieve the next set of results.

$sel:pipelineId:QueryObjects', queryObjects_pipelineId - The ID of the pipeline.

$sel:sphere:QueryObjects', queryObjects_sphere - Indicates whether the query applies to components or instances. The possible values are: COMPONENT, INSTANCE, and ATTEMPT.

data QueryObjectsResponse

Contains the output of QueryObjects.

See: newQueryObjectsResponse smart constructor.

Constructors

QueryObjectsResponse' (Maybe [Text]) (Maybe Bool) (Maybe Text) Int

newQueryObjectsResponse

Arguments

:: Int
$sel:httpStatus:QueryObjectsResponse'

-> QueryObjectsResponse

Create a value of QueryObjectsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ids:QueryObjectsResponse', queryObjectsResponse_ids - The identifiers that match the query selectors.

$sel:hasMoreResults:QueryObjectsResponse', queryObjectsResponse_hasMoreResults - Indicates whether there are more results that can be obtained by a subsequent call.

$sel:marker:QueryObjects', queryObjectsResponse_marker - The starting point for the next page of results. To view the next page of results, call QueryObjects again with this marker value. If the value is null, there are no more results.

$sel:httpStatus:QueryObjectsResponse', queryObjectsResponse_httpStatus - The response's http status code.

EvaluateExpression

data EvaluateExpression

Contains the parameters for EvaluateExpression.

See: newEvaluateExpression smart constructor.

Constructors

EvaluateExpression' Text Text Text

newEvaluateExpression

Arguments

:: Text
$sel:pipelineId:EvaluateExpression'

-> Text
$sel:objectId:EvaluateExpression'

-> Text
$sel:expression:EvaluateExpression'

-> EvaluateExpression

Create a value of EvaluateExpression with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pipelineId:EvaluateExpression', evaluateExpression_pipelineId - The ID of the pipeline.

$sel:objectId:EvaluateExpression', evaluateExpression_objectId - The ID of the object.

$sel:expression:EvaluateExpression', evaluateExpression_expression - The expression to evaluate.

data EvaluateExpressionResponse

Contains the output of EvaluateExpression.

See: newEvaluateExpressionResponse smart constructor.

Constructors

EvaluateExpressionResponse' Int Text

newEvaluateExpressionResponse

Arguments

:: Int
$sel:httpStatus:EvaluateExpressionResponse'

-> Text
$sel:evaluatedExpression:EvaluateExpressionResponse'

-> EvaluateExpressionResponse

Create a value of EvaluateExpressionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:EvaluateExpressionResponse', evaluateExpressionResponse_httpStatus - The response's http status code.

$sel:evaluatedExpression:EvaluateExpressionResponse', evaluateExpressionResponse_evaluatedExpression - The evaluated expression.

ReportTaskProgress

data ReportTaskProgress

Contains the parameters for ReportTaskProgress.

See: newReportTaskProgress smart constructor.

Constructors

ReportTaskProgress' (Maybe [Field]) Text

newReportTaskProgress

Arguments

:: Text
$sel:taskId:ReportTaskProgress'

-> ReportTaskProgress

Create a value of ReportTaskProgress with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fields:ReportTaskProgress', reportTaskProgress_fields - Key-value pairs that define the properties of the ReportTaskProgressInput object.

$sel:taskId:ReportTaskProgress', reportTaskProgress_taskId - The ID of the task assigned to the task runner. This value is provided in the response for PollForTask.

data ReportTaskProgressResponse

Contains the output of ReportTaskProgress.

See: newReportTaskProgressResponse smart constructor.

Constructors

ReportTaskProgressResponse' Int Bool

newReportTaskProgressResponse

Arguments

:: Int
$sel:httpStatus:ReportTaskProgressResponse'

-> Bool
$sel:canceled:ReportTaskProgressResponse'

-> ReportTaskProgressResponse

Create a value of ReportTaskProgressResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:ReportTaskProgressResponse', reportTaskProgressResponse_httpStatus - The response's http status code.

$sel:canceled:ReportTaskProgressResponse', reportTaskProgressResponse_canceled - If true, the calling task runner should cancel processing of the task. The task runner does not need to call SetTaskStatus for canceled tasks.

CreatePipeline

data CreatePipeline

Contains the parameters for CreatePipeline.

See: newCreatePipeline smart constructor.

Constructors

CreatePipeline' (Maybe [Tag]) (Maybe Text) Text Text

newCreatePipeline

Arguments

:: Text
$sel:name:CreatePipeline'

-> Text
$sel:uniqueId:CreatePipeline'

-> CreatePipeline

Create a value of CreatePipeline with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreatePipeline', createPipeline_tags - A list of tags to associate with the pipeline at creation. Tags let you control access to pipelines. For more information, see Controlling User Access to Pipelines in the AWS Data Pipeline Developer Guide.

$sel:description:CreatePipeline', createPipeline_description - The description for the pipeline.

$sel:name:CreatePipeline', createPipeline_name - The name for the pipeline. You can use the same name for multiple pipelines associated with your AWS account, because AWS Data Pipeline assigns each pipeline a unique pipeline identifier.

$sel:uniqueId:CreatePipeline', createPipeline_uniqueId - A unique identifier. This identifier is not the same as the pipeline identifier assigned by AWS Data Pipeline. You are responsible for defining the format and ensuring the uniqueness of this identifier. You use this parameter to ensure idempotency during repeated calls to CreatePipeline. For example, if the first call to CreatePipeline does not succeed, you can pass in the same unique identifier and pipeline name combination on a subsequent call to CreatePipeline. CreatePipeline ensures that if a pipeline already exists with the same name and unique identifier, a new pipeline is not created. Instead, you'll receive the pipeline identifier from the previous attempt. The uniqueness of the name and unique identifier combination is scoped to the AWS account or IAM user credentials.

data CreatePipelineResponse

Contains the output of CreatePipeline.

See: newCreatePipelineResponse smart constructor.

Constructors

CreatePipelineResponse' Int Text

newCreatePipelineResponse

Arguments

:: Int
$sel:httpStatus:CreatePipelineResponse'

-> Text
$sel:pipelineId:CreatePipelineResponse'

-> CreatePipelineResponse

Create a value of CreatePipelineResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreatePipelineResponse', createPipelineResponse_httpStatus - The response's http status code.

$sel:pipelineId:CreatePipelineResponse', createPipelineResponse_pipelineId - The ID that AWS Data Pipeline assigns the newly created pipeline. For example, df-06372391ZG65EXAMPLE.

DeletePipeline

data DeletePipeline

Contains the parameters for DeletePipeline.

See: newDeletePipeline smart constructor.

Constructors

DeletePipeline' Text

newDeletePipeline

Arguments

:: Text
$sel:pipelineId:DeletePipeline'

-> DeletePipeline

Create a value of DeletePipeline with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pipelineId:DeletePipeline', deletePipeline_pipelineId - The ID of the pipeline.

data DeletePipelineResponse

See: newDeletePipelineResponse smart constructor.

Constructors

DeletePipelineResponse'

newDeletePipelineResponse :: DeletePipelineResponse

Create a value of DeletePipelineResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

SetTaskStatus

data SetTaskStatus

Contains the parameters for SetTaskStatus.

See: newSetTaskStatus smart constructor.

Constructors

SetTaskStatus' (Maybe Text) (Maybe Text) (Maybe Text) Text TaskStatus

newSetTaskStatus

Arguments

:: Text
$sel:taskId:SetTaskStatus'

-> TaskStatus
$sel:taskStatus:SetTaskStatus'

-> SetTaskStatus

Create a value of SetTaskStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:errorStackTrace:SetTaskStatus', setTaskStatus_errorStackTrace - If an error occurred during the task, this value specifies the stack trace associated with the error. This value is set on the physical attempt object. It is used to display error information to the user. The web service does not parse this value.

$sel:errorMessage:SetTaskStatus', setTaskStatus_errorMessage - If an error occurred during the task, this value specifies a text description of the error. This value is set on the physical attempt object. It is used to display error information to the user. The web service does not parse this value.

$sel:errorId:SetTaskStatus', setTaskStatus_errorId - If an error occurred during the task, this value specifies the error code. This value is set on the physical attempt object. It is used to display error information to the user. It should not start with string "Service_" which is reserved by the system.

$sel:taskId:SetTaskStatus', setTaskStatus_taskId - The ID of the task assigned to the task runner. This value is provided in the response for PollForTask.

$sel:taskStatus:SetTaskStatus', setTaskStatus_taskStatus - If FINISHED, the task successfully completed. If FAILED, the task ended unsuccessfully. Preconditions use false.

data SetTaskStatusResponse

Contains the output of SetTaskStatus.

See: newSetTaskStatusResponse smart constructor.

Constructors

SetTaskStatusResponse' Int

newSetTaskStatusResponse

Arguments

:: Int
$sel:httpStatus:SetTaskStatusResponse'

-> SetTaskStatusResponse

Create a value of SetTaskStatusResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:SetTaskStatusResponse', setTaskStatusResponse_httpStatus - The response's http status code.

ReportTaskRunnerHeartbeat

data ReportTaskRunnerHeartbeat

Contains the parameters for ReportTaskRunnerHeartbeat.

See: newReportTaskRunnerHeartbeat smart constructor.

Constructors

ReportTaskRunnerHeartbeat' (Maybe Text) (Maybe Text) Text

newReportTaskRunnerHeartbeat

Arguments

:: Text
$sel:taskrunnerId:ReportTaskRunnerHeartbeat'

-> ReportTaskRunnerHeartbeat

Create a value of ReportTaskRunnerHeartbeat with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hostname:ReportTaskRunnerHeartbeat', reportTaskRunnerHeartbeat_hostname - The public DNS name of the task runner.

$sel:workerGroup:ReportTaskRunnerHeartbeat', reportTaskRunnerHeartbeat_workerGroup - The type of task the task runner is configured to accept and process. The worker group is set as a field on objects in the pipeline when they are created. You can only specify a single value for workerGroup. There are no wildcard values permitted in workerGroup; the string must be an exact, case-sensitive, match.

$sel:taskrunnerId:ReportTaskRunnerHeartbeat', reportTaskRunnerHeartbeat_taskrunnerId - The ID of the task runner. This value should be unique across your AWS account. In the case of AWS Data Pipeline Task Runner launched on a resource managed by AWS Data Pipeline, the web service provides a unique identifier when it launches the application. If you have written a custom task runner, you should assign a unique identifier for the task runner.

data ReportTaskRunnerHeartbeatResponse

Contains the output of ReportTaskRunnerHeartbeat.

See: newReportTaskRunnerHeartbeatResponse smart constructor.

Constructors

ReportTaskRunnerHeartbeatResponse' Int Bool

newReportTaskRunnerHeartbeatResponse

Arguments

:: Int
$sel:httpStatus:ReportTaskRunnerHeartbeatResponse'

-> Bool
$sel:terminate:ReportTaskRunnerHeartbeatResponse'

-> ReportTaskRunnerHeartbeatResponse

Create a value of ReportTaskRunnerHeartbeatResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:ReportTaskRunnerHeartbeatResponse', reportTaskRunnerHeartbeatResponse_httpStatus - The response's http status code.

$sel:terminate:ReportTaskRunnerHeartbeatResponse', reportTaskRunnerHeartbeatResponse_terminate - Indicates whether the calling task runner should terminate.

DescribeObjects (Paginated)

data DescribeObjects

Contains the parameters for DescribeObjects.

See: newDescribeObjects smart constructor.

Constructors

DescribeObjects' (Maybe Bool) (Maybe Text) Text [Text]

newDescribeObjects

Arguments

:: Text
$sel:pipelineId:DescribeObjects'

-> DescribeObjects

Create a value of DescribeObjects with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:evaluateExpressions:DescribeObjects', describeObjects_evaluateExpressions - Indicates whether any expressions in the object should be evaluated when the object descriptions are returned.

$sel:marker:DescribeObjects', describeObjects_marker - The starting point for the results to be returned. For the first call, this value should be empty. As long as there are more results, continue to call DescribeObjects with the marker value from the previous call to retrieve the next set of results.

$sel:pipelineId:DescribeObjects', describeObjects_pipelineId - The ID of the pipeline that contains the object definitions.

$sel:objectIds:DescribeObjects', describeObjects_objectIds - The IDs of the pipeline objects that contain the definitions to be described. You can pass as many as 25 identifiers in a single call to DescribeObjects.

data DescribeObjectsResponse

Contains the output of DescribeObjects.

See: newDescribeObjectsResponse smart constructor.

Constructors

DescribeObjectsResponse' (Maybe Bool) (Maybe Text) Int [PipelineObject]

newDescribeObjectsResponse

Arguments

:: Int
$sel:httpStatus:DescribeObjectsResponse'

-> DescribeObjectsResponse

Create a value of DescribeObjectsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hasMoreResults:DescribeObjectsResponse', describeObjectsResponse_hasMoreResults - Indicates whether there are more results to return.

$sel:marker:DescribeObjects', describeObjectsResponse_marker - The starting point for the next page of results. To view the next page of results, call DescribeObjects again with this marker value. If the value is null, there are no more results.

$sel:httpStatus:DescribeObjectsResponse', describeObjectsResponse_httpStatus - The response's http status code.

$sel:pipelineObjects:DescribeObjectsResponse', describeObjectsResponse_pipelineObjects - An array of object definitions.

AddTags

data AddTags

Contains the parameters for AddTags.

See: newAddTags smart constructor.

Constructors

AddTags' Text [Tag]

newAddTags

Arguments

:: Text
$sel:pipelineId:AddTags'

-> AddTags

Create a value of AddTags with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pipelineId:AddTags', addTags_pipelineId - The ID of the pipeline.

$sel:tags:AddTags', addTags_tags - The tags to add, as key/value pairs.

data AddTagsResponse

Contains the output of AddTags.

See: newAddTagsResponse smart constructor.

Constructors

AddTagsResponse' Int

newAddTagsResponse

Arguments

:: Int
$sel:httpStatus:AddTagsResponse'

-> AddTagsResponse

Create a value of AddTagsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:AddTagsResponse', addTagsResponse_httpStatus - The response's http status code.

DescribePipelines

data DescribePipelines

Contains the parameters for DescribePipelines.

See: newDescribePipelines smart constructor.

Constructors

DescribePipelines' [Text]

newDescribePipelines :: DescribePipelines

Create a value of DescribePipelines with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pipelineIds:DescribePipelines', describePipelines_pipelineIds - The IDs of the pipelines to describe. You can pass as many as 25 identifiers in a single call. To obtain pipeline IDs, call ListPipelines.

data DescribePipelinesResponse

Contains the output of DescribePipelines.

See: newDescribePipelinesResponse smart constructor.

Constructors

DescribePipelinesResponse' Int [PipelineDescription]

newDescribePipelinesResponse

Arguments

:: Int
$sel:httpStatus:DescribePipelinesResponse'

-> DescribePipelinesResponse

Create a value of DescribePipelinesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DescribePipelinesResponse', describePipelinesResponse_httpStatus - The response's http status code.

$sel:pipelineDescriptionList:DescribePipelinesResponse', describePipelinesResponse_pipelineDescriptionList - An array of descriptions for the specified pipelines.

PollForTask

data PollForTask

Contains the parameters for PollForTask.

See: newPollForTask smart constructor.

Constructors

PollForTask' (Maybe Text) (Maybe InstanceIdentity) Text

newPollForTask

Arguments

:: Text
$sel:workerGroup:PollForTask'

-> PollForTask

Create a value of PollForTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hostname:PollForTask', pollForTask_hostname - The public DNS name of the calling task runner.

$sel:instanceIdentity:PollForTask', pollForTask_instanceIdentity - Identity information for the EC2 instance that is hosting the task runner. You can get this value from the instance using http://169.254.169.254/latest/meta-data/instance-id. For more information, see Instance Metadata in the Amazon Elastic Compute Cloud User Guide. Passing in this value proves that your task runner is running on an EC2 instance, and ensures the proper AWS Data Pipeline service charges are applied to your pipeline.

$sel:workerGroup:PollForTask', pollForTask_workerGroup - The type of task the task runner is configured to accept and process. The worker group is set as a field on objects in the pipeline when they are created. You can only specify a single value for workerGroup in the call to PollForTask. There are no wildcard values permitted in workerGroup; the string must be an exact, case-sensitive, match.

data PollForTaskResponse

Contains the output of PollForTask.

See: newPollForTaskResponse smart constructor.

Constructors

PollForTaskResponse' (Maybe TaskObject) Int

newPollForTaskResponse

Arguments

:: Int
$sel:httpStatus:PollForTaskResponse'

-> PollForTaskResponse

Create a value of PollForTaskResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:taskObject:PollForTaskResponse', pollForTaskResponse_taskObject - The information needed to complete the task that is being assigned to the task runner. One of the fields returned in this object is taskId, which contains an identifier for the task being assigned. The calling task runner uses taskId in subsequent calls to ReportTaskProgress and SetTaskStatus.

$sel:httpStatus:PollForTaskResponse', pollForTaskResponse_httpStatus - The response's http status code.

ValidatePipelineDefinition

data ValidatePipelineDefinition

Contains the parameters for ValidatePipelineDefinition.

See: newValidatePipelineDefinition smart constructor.

Constructors

ValidatePipelineDefinition' (Maybe [ParameterValue]) (Maybe [ParameterObject]) Text [PipelineObject]

newValidatePipelineDefinition

Arguments

:: Text
$sel:pipelineId:ValidatePipelineDefinition'

-> ValidatePipelineDefinition

Create a value of ValidatePipelineDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parameterValues:ValidatePipelineDefinition', validatePipelineDefinition_parameterValues - The parameter values used with the pipeline.

$sel:parameterObjects:ValidatePipelineDefinition', validatePipelineDefinition_parameterObjects - The parameter objects used with the pipeline.

$sel:pipelineId:ValidatePipelineDefinition', validatePipelineDefinition_pipelineId - The ID of the pipeline.

$sel:pipelineObjects:ValidatePipelineDefinition', validatePipelineDefinition_pipelineObjects - The objects that define the pipeline changes to validate against the pipeline.

data ValidatePipelineDefinitionResponse

Contains the output of ValidatePipelineDefinition.

See: newValidatePipelineDefinitionResponse smart constructor.

Constructors

ValidatePipelineDefinitionResponse' (Maybe [ValidationError]) (Maybe [ValidationWarning]) Int Bool

newValidatePipelineDefinitionResponse

Arguments

:: Int
$sel:httpStatus:ValidatePipelineDefinitionResponse'

-> Bool
$sel:errored:ValidatePipelineDefinitionResponse'

-> ValidatePipelineDefinitionResponse

Create a value of ValidatePipelineDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:validationErrors:ValidatePipelineDefinitionResponse', validatePipelineDefinitionResponse_validationErrors - Any validation errors that were found.

$sel:validationWarnings:ValidatePipelineDefinitionResponse', validatePipelineDefinitionResponse_validationWarnings - Any validation warnings that were found.

$sel:httpStatus:ValidatePipelineDefinitionResponse', validatePipelineDefinitionResponse_httpStatus - The response's http status code.

$sel:errored:ValidatePipelineDefinitionResponse', validatePipelineDefinitionResponse_errored - Indicates whether there were validation errors.

GetPipelineDefinition

data GetPipelineDefinition

Contains the parameters for GetPipelineDefinition.

See: newGetPipelineDefinition smart constructor.

Constructors

GetPipelineDefinition' (Maybe Text) Text

newGetPipelineDefinition

Arguments

:: Text
$sel:pipelineId:GetPipelineDefinition'

-> GetPipelineDefinition

Create a value of GetPipelineDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:version:GetPipelineDefinition', getPipelineDefinition_version - The version of the pipeline definition to retrieve. Set this parameter to latest (default) to use the last definition saved to the pipeline or active to use the last definition that was activated.

$sel:pipelineId:GetPipelineDefinition', getPipelineDefinition_pipelineId - The ID of the pipeline.

data GetPipelineDefinitionResponse

Contains the output of GetPipelineDefinition.

See: newGetPipelineDefinitionResponse smart constructor.

Constructors

GetPipelineDefinitionResponse' (Maybe [ParameterValue]) (Maybe [ParameterObject]) (Maybe [PipelineObject]) Int

newGetPipelineDefinitionResponse

Arguments

:: Int
$sel:httpStatus:GetPipelineDefinitionResponse'

-> GetPipelineDefinitionResponse

Create a value of GetPipelineDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parameterValues:GetPipelineDefinitionResponse', getPipelineDefinitionResponse_parameterValues - The parameter values used in the pipeline definition.

$sel:parameterObjects:GetPipelineDefinitionResponse', getPipelineDefinitionResponse_parameterObjects - The parameter objects used in the pipeline definition.

$sel:pipelineObjects:GetPipelineDefinitionResponse', getPipelineDefinitionResponse_pipelineObjects - The objects defined in the pipeline.

$sel:httpStatus:GetPipelineDefinitionResponse', getPipelineDefinitionResponse_httpStatus - The response's http status code.

PutPipelineDefinition

data PutPipelineDefinition

Contains the parameters for PutPipelineDefinition.

See: newPutPipelineDefinition smart constructor.

Constructors

PutPipelineDefinition' (Maybe [ParameterValue]) (Maybe [ParameterObject]) Text [PipelineObject]

newPutPipelineDefinition

Arguments

:: Text
$sel:pipelineId:PutPipelineDefinition'

-> PutPipelineDefinition

Create a value of PutPipelineDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parameterValues:PutPipelineDefinition', putPipelineDefinition_parameterValues - The parameter values used with the pipeline.

$sel:parameterObjects:PutPipelineDefinition', putPipelineDefinition_parameterObjects - The parameter objects used with the pipeline.

$sel:pipelineId:PutPipelineDefinition', putPipelineDefinition_pipelineId - The ID of the pipeline.

$sel:pipelineObjects:PutPipelineDefinition', putPipelineDefinition_pipelineObjects - The objects that define the pipeline. These objects overwrite the existing pipeline definition.

data PutPipelineDefinitionResponse

Contains the output of PutPipelineDefinition.

See: newPutPipelineDefinitionResponse smart constructor.

Constructors

PutPipelineDefinitionResponse' (Maybe [ValidationError]) (Maybe [ValidationWarning]) Int Bool

newPutPipelineDefinitionResponse

Arguments

:: Int
$sel:httpStatus:PutPipelineDefinitionResponse'

-> Bool
$sel:errored:PutPipelineDefinitionResponse'

-> PutPipelineDefinitionResponse

Create a value of PutPipelineDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:validationErrors:PutPipelineDefinitionResponse', putPipelineDefinitionResponse_validationErrors - The validation errors that are associated with the objects defined in pipelineObjects.

$sel:validationWarnings:PutPipelineDefinitionResponse', putPipelineDefinitionResponse_validationWarnings - The validation warnings that are associated with the objects defined in pipelineObjects.

$sel:httpStatus:PutPipelineDefinitionResponse', putPipelineDefinitionResponse_httpStatus - The response's http status code.

$sel:errored:PutPipelineDefinitionResponse', putPipelineDefinitionResponse_errored - Indicates whether there were validation errors, and the pipeline definition is stored but cannot be activated until you correct the pipeline and call PutPipelineDefinition to commit the corrected pipeline.

SetStatus

data SetStatus

Contains the parameters for SetStatus.

See: newSetStatus smart constructor.

Constructors

SetStatus' Text [Text] Text

newSetStatus

Arguments

:: Text
$sel:pipelineId:SetStatus'

-> Text
$sel:status:SetStatus'

-> SetStatus

Create a value of SetStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pipelineId:SetStatus', setStatus_pipelineId - The ID of the pipeline that contains the objects.

$sel:objectIds:SetStatus', setStatus_objectIds - The IDs of the objects. The corresponding objects can be either physical or components, but not a mix of both types.

$sel:status:SetStatus', setStatus_status - The status to be set on all the objects specified in objectIds. For components, use PAUSE or RESUME. For instances, use TRY_CANCEL, RERUN, or MARK_FINISHED.

data SetStatusResponse

See: newSetStatusResponse smart constructor.

Constructors

SetStatusResponse'

newSetStatusResponse :: SetStatusResponse

Create a value of SetStatusResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

ListPipelines (Paginated)

data ListPipelines

Contains the parameters for ListPipelines.

See: newListPipelines smart constructor.

Constructors

ListPipelines' (Maybe Text)

newListPipelines :: ListPipelines

Create a value of ListPipelines with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:marker:ListPipelines', listPipelines_marker - The starting point for the results to be returned. For the first call, this value should be empty. As long as there are more results, continue to call ListPipelines with the marker value from the previous call to retrieve the next set of results.

data ListPipelinesResponse

Contains the output of ListPipelines.

See: newListPipelinesResponse smart constructor.

Constructors

ListPipelinesResponse' (Maybe Bool) (Maybe Text) Int [PipelineIdName]

newListPipelinesResponse

Arguments

:: Int
$sel:httpStatus:ListPipelinesResponse'

-> ListPipelinesResponse

Create a value of ListPipelinesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hasMoreResults:ListPipelinesResponse', listPipelinesResponse_hasMoreResults - Indicates whether there are more results that can be obtained by a subsequent call.

$sel:marker:ListPipelines', listPipelinesResponse_marker - The starting point for the next page of results. To view the next page of results, call ListPipelinesOutput again with this marker value. If the value is null, there are no more results.

$sel:httpStatus:ListPipelinesResponse', listPipelinesResponse_httpStatus - The response's http status code.

$sel:pipelineIdList:ListPipelinesResponse', listPipelinesResponse_pipelineIdList - The pipeline identifiers. If you require additional information about the pipelines, you can use these identifiers to call DescribePipelines and GetPipelineDefinition.

ActivatePipeline

data ActivatePipeline

Contains the parameters for ActivatePipeline.

See: newActivatePipeline smart constructor.

Constructors

ActivatePipeline' (Maybe [ParameterValue]) (Maybe POSIX) Text

newActivatePipeline

Arguments

:: Text
$sel:pipelineId:ActivatePipeline'

-> ActivatePipeline

Create a value of ActivatePipeline with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parameterValues:ActivatePipeline', activatePipeline_parameterValues - A list of parameter values to pass to the pipeline at activation.

$sel:startTimestamp:ActivatePipeline', activatePipeline_startTimestamp - The date and time to resume the pipeline. By default, the pipeline resumes from the last completed execution.

$sel:pipelineId:ActivatePipeline', activatePipeline_pipelineId - The ID of the pipeline.

data ActivatePipelineResponse

Contains the output of ActivatePipeline.

See: newActivatePipelineResponse smart constructor.

Constructors

ActivatePipelineResponse' Int

newActivatePipelineResponse

Arguments

:: Int
$sel:httpStatus:ActivatePipelineResponse'

-> ActivatePipelineResponse

Create a value of ActivatePipelineResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:ActivatePipelineResponse', activatePipelineResponse_httpStatus - The response's http status code.

DeactivatePipeline

data DeactivatePipeline

Contains the parameters for DeactivatePipeline.

See: newDeactivatePipeline smart constructor.

Constructors

DeactivatePipeline' (Maybe Bool) Text

newDeactivatePipeline

Arguments

:: Text
$sel:pipelineId:DeactivatePipeline'

-> DeactivatePipeline

Create a value of DeactivatePipeline with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cancelActive:DeactivatePipeline', deactivatePipeline_cancelActive - Indicates whether to cancel any running objects. The default is true, which sets the state of any running objects to CANCELED. If this value is false, the pipeline is deactivated after all running objects finish.

$sel:pipelineId:DeactivatePipeline', deactivatePipeline_pipelineId - The ID of the pipeline.

data DeactivatePipelineResponse

Contains the output of DeactivatePipeline.

See: newDeactivatePipelineResponse smart constructor.

Constructors

DeactivatePipelineResponse' Int

newDeactivatePipelineResponse

Arguments

:: Int
$sel:httpStatus:DeactivatePipelineResponse'

-> DeactivatePipelineResponse

Create a value of DeactivatePipelineResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeactivatePipelineResponse', deactivatePipelineResponse_httpStatus - The response's http status code.

Types

OperatorType

newtype OperatorType

Constructors

OperatorType'

Fields

fromOperatorType :: Text

Bundled Patterns

pattern OperatorType_BETWEEN :: OperatorType

pattern OperatorType_EQ :: OperatorType

pattern OperatorType_GE :: OperatorType

pattern OperatorType_LE :: OperatorType

pattern OperatorType_REF_EQ :: OperatorType

TaskStatus

newtype TaskStatus

Constructors

TaskStatus'

Fields

fromTaskStatus :: Text

Bundled Patterns

pattern TaskStatus_FAILED :: TaskStatus

pattern TaskStatus_FALSE :: TaskStatus

pattern TaskStatus_FINISHED :: TaskStatus

Field

data Field

A key-value pair that describes a property of a pipeline object. The value is specified as either a string value (StringValue) or a reference to another object (RefValue) but not as both.

See: newField smart constructor.

Constructors

Field' (Maybe Text) (Maybe Text) Text

newField

Arguments

:: Text
$sel:key:Field'

-> Field

Create a value of Field with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stringValue:Field', field_stringValue - The field value, expressed as a String.

$sel:refValue:Field', field_refValue - The field value, expressed as the identifier of another object.

$sel:key:Field', field_key - The field identifier.

InstanceIdentity

data InstanceIdentity

Identity information for the EC2 instance that is hosting the task runner. You can get this value by calling a metadata URI from the EC2 instance. For more information, see Instance Metadata in the Amazon Elastic Compute Cloud User Guide. Passing in this value proves that your task runner is running on an EC2 instance, and ensures the proper AWS Data Pipeline service charges are applied to your pipeline.

See: newInstanceIdentity smart constructor.

Constructors

InstanceIdentity' (Maybe Text) (Maybe Text)

newInstanceIdentity :: InstanceIdentity

Create a value of InstanceIdentity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:document:InstanceIdentity', instanceIdentity_document - A description of an EC2 instance that is generated when the instance is launched and exposed to the instance via the instance metadata service in the form of a JSON representation of an object.

$sel:signature:InstanceIdentity', instanceIdentity_signature - A signature which can be used to verify the accuracy and authenticity of the information provided in the instance identity document.

Operator

data Operator

Contains a logical operation for comparing the value of a field with a specified value.

See: newOperator smart constructor.

Constructors

Operator' (Maybe [Text]) (Maybe OperatorType)

newOperator :: Operator

Create a value of Operator with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:values:Operator', operator_values - The value that the actual field value will be compared with.

$sel:type':Operator', operator_type - The logical operation to be performed: equal (EQ), equal reference (REF_EQ), less than or equal (LE), greater than or equal (GE), or between (BETWEEN). Equal reference (REF_EQ) can be used only with reference fields. The other comparison types can be used only with String fields. The comparison types you can use apply only to certain object fields, as detailed below.

The comparison operators EQ and REF_EQ act on the following fields:

name

@sphere

parent

@componentParent

@instanceParent

@status

@scheduledStartTime

@scheduledEndTime

@actualStartTime

@actualEndTime

The comparison operators GE, LE, and BETWEEN act on the following fields:

@scheduledStartTime

@scheduledEndTime

@actualStartTime

@actualEndTime

Note that fields beginning with the at sign (@) are read-only and set by the web service. When you name fields, you should choose names containing only alpha-numeric values, as symbols may be reserved by AWS Data Pipeline. User-defined fields that you add to a pipeline should prefix their name with the string "my".

ParameterAttribute

data ParameterAttribute

The attributes allowed or specified with a parameter object.

See: newParameterAttribute smart constructor.

Constructors

ParameterAttribute' Text Text

newParameterAttribute

Arguments

:: Text
$sel:key:ParameterAttribute'

-> Text
$sel:stringValue:ParameterAttribute'

-> ParameterAttribute

Create a value of ParameterAttribute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:ParameterAttribute', parameterAttribute_key - The field identifier.

$sel:stringValue:ParameterAttribute', parameterAttribute_stringValue - The field value, expressed as a String.

ParameterObject

data ParameterObject

Contains information about a parameter object.

See: newParameterObject smart constructor.

Constructors

ParameterObject' Text [ParameterAttribute]

newParameterObject

Arguments

:: Text
$sel:id:ParameterObject'

-> ParameterObject

Create a value of ParameterObject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:ParameterObject', parameterObject_id - The ID of the parameter object.

$sel:attributes:ParameterObject', parameterObject_attributes - The attributes of the parameter object.

ParameterValue

data ParameterValue

A value or list of parameter values.

See: newParameterValue smart constructor.

Constructors

ParameterValue' Text Text

newParameterValue

Arguments

:: Text
$sel:id:ParameterValue'

-> Text
$sel:stringValue:ParameterValue'

-> ParameterValue

Create a value of ParameterValue with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:ParameterValue', parameterValue_id - The ID of the parameter value.

$sel:stringValue:ParameterValue', parameterValue_stringValue - The field value, expressed as a String.

PipelineDescription

data PipelineDescription

Contains pipeline metadata.

See: newPipelineDescription smart constructor.

Constructors

PipelineDescription' (Maybe [Tag]) (Maybe Text) Text Text [Field]

newPipelineDescription

Arguments

:: Text
$sel:pipelineId:PipelineDescription'

-> Text
$sel:name:PipelineDescription'

-> PipelineDescription

Create a value of PipelineDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:PipelineDescription', pipelineDescription_tags - A list of tags to associated with a pipeline. Tags let you control access to pipelines. For more information, see Controlling User Access to Pipelines in the AWS Data Pipeline Developer Guide.

$sel:description:PipelineDescription', pipelineDescription_description - Description of the pipeline.

$sel:pipelineId:PipelineDescription', pipelineDescription_pipelineId - The pipeline identifier that was assigned by AWS Data Pipeline. This is a string of the form df-297EG78HU43EEXAMPLE.

$sel:name:PipelineDescription', pipelineDescription_name - The name of the pipeline.

$sel:fields:PipelineDescription', pipelineDescription_fields - A list of read-only fields that contain metadata about the pipeline: @userId, @accountId, and @pipelineState.

PipelineIdName

data PipelineIdName

Contains the name and identifier of a pipeline.

See: newPipelineIdName smart constructor.

Constructors

PipelineIdName' (Maybe Text) (Maybe Text)

newPipelineIdName :: PipelineIdName

Create a value of PipelineIdName with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:PipelineIdName', pipelineIdName_id - The ID of the pipeline that was assigned by AWS Data Pipeline. This is a string of the form df-297EG78HU43EEXAMPLE.

$sel:name:PipelineIdName', pipelineIdName_name - The name of the pipeline.

PipelineObject

data PipelineObject

Contains information about a pipeline object. This can be a logical, physical, or physical attempt pipeline object. The complete set of components of a pipeline defines the pipeline.

See: newPipelineObject smart constructor.

Constructors

PipelineObject' Text Text [Field]

newPipelineObject

Arguments

:: Text
$sel:id:PipelineObject'

-> Text
$sel:name:PipelineObject'

-> PipelineObject

Create a value of PipelineObject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:PipelineObject', pipelineObject_id - The ID of the object.

$sel:name:PipelineObject', pipelineObject_name - The name of the object.

$sel:fields:PipelineObject', pipelineObject_fields - Key-value pairs that define the properties of the object.

Query

data Query

Defines the query to run against an object.

See: newQuery smart constructor.

Constructors

Query' (Maybe [Selector])

newQuery :: Query

Create a value of Query with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:selectors:Query', query_selectors - List of selectors that define the query. An object must satisfy all of the selectors to match the query.

Selector

data Selector

A comparision that is used to determine whether a query should return this object.

See: newSelector smart constructor.

Constructors

Selector' (Maybe Operator) (Maybe Text)

newSelector :: Selector

Create a value of Selector with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operator:Selector', selector_operator - Undocumented member.

$sel:fieldName:Selector', selector_fieldName - The name of the field that the operator will be applied to. The field name is the "key" portion of the field definition in the pipeline definition syntax that is used by the AWS Data Pipeline API. If the field is not set on the object, the condition fails.

Tag

data Tag

Tags are key/value pairs defined by a user and associated with a pipeline to control access. AWS Data Pipeline allows you to associate ten tags per pipeline. For more information, see Controlling User Access to Pipelines in the AWS Data Pipeline Developer Guide.

See: newTag smart constructor.

Constructors

Tag' Text Text

newTag

Arguments

:: Text
$sel:key:Tag'

-> Text
$sel:value:Tag'

-> Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:Tag', tag_key - The key name of a tag defined by a user. For more information, see Controlling User Access to Pipelines in the AWS Data Pipeline Developer Guide.

$sel:value:Tag', tag_value - The optional value portion of a tag defined by a user. For more information, see Controlling User Access to Pipelines in the AWS Data Pipeline Developer Guide.

TaskObject

data TaskObject

Contains information about a pipeline task that is assigned to a task runner.

See: newTaskObject smart constructor.

Constructors

TaskObject' (Maybe Text) (Maybe (HashMap Text PipelineObject)) (Maybe Text) (Maybe Text)

newTaskObject :: TaskObject

Create a value of TaskObject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pipelineId:TaskObject', taskObject_pipelineId - The ID of the pipeline that provided the task.

$sel:objects:TaskObject', taskObject_objects - Connection information for the location where the task runner will publish the output of the task.

$sel:taskId:TaskObject', taskObject_taskId - An internal identifier for the task. This ID is passed to the SetTaskStatus and ReportTaskProgress actions.

$sel:attemptId:TaskObject', taskObject_attemptId - The ID of the pipeline task attempt object. AWS Data Pipeline uses this value to track how many times a task is attempted.

ValidationError

data ValidationError

Defines a validation error. Validation errors prevent pipeline activation. The set of validation errors that can be returned are defined by AWS Data Pipeline.

See: newValidationError smart constructor.

Constructors

ValidationError' (Maybe Text) (Maybe [Text])

newValidationError :: ValidationError

Create a value of ValidationError with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:ValidationError', validationError_id - The identifier of the object that contains the validation error.

$sel:errors:ValidationError', validationError_errors - A description of the validation error.

ValidationWarning

data ValidationWarning

Defines a validation warning. Validation warnings do not prevent pipeline activation. The set of validation warnings that can be returned are defined by AWS Data Pipeline.

See: newValidationWarning smart constructor.

Constructors

ValidationWarning' (Maybe [Text]) (Maybe Text)

newValidationWarning :: ValidationWarning

Create a value of ValidationWarning with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:warnings:ValidationWarning', validationWarning_warnings - A description of the validation warning.

$sel:id:ValidationWarning', validationWarning_id - The identifier of the object that contains the validation warning.