Network.AWS.DataPipeline.DescribeObjects

Description

Gets the object definitions for a set of objects associated with the pipeline. Object definitions are composed of a set of fields that define the properties of the object.

This operation returns paginated results.

Creating a Request

data DescribeObjects

Contains the parameters for DescribeObjects.

See: newDescribeObjects smart constructor.

Constructors

DescribeObjects'

Fields

evaluateExpressions :: Maybe Bool

Indicates whether any expressions in the object should be evaluated when the object descriptions are returned.

marker :: Maybe Text

The starting point for the results to be returned. For the first call, this value should be empty. As long as there are more results, continue to call DescribeObjects with the marker value from the previous call to retrieve the next set of results.

pipelineId :: Text

The ID of the pipeline that contains the object definitions.

objectIds :: [Text]

The IDs of the pipeline objects that contain the definitions to be described. You can pass as many as 25 identifiers in a single call to DescribeObjects.

newDescribeObjects

Arguments

:: Text
$sel:pipelineId:DescribeObjects'

-> DescribeObjects

Create a value of DescribeObjects with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:evaluateExpressions:DescribeObjects', describeObjects_evaluateExpressions - Indicates whether any expressions in the object should be evaluated when the object descriptions are returned.

$sel:marker:DescribeObjects', describeObjects_marker - The starting point for the results to be returned. For the first call, this value should be empty. As long as there are more results, continue to call DescribeObjects with the marker value from the previous call to retrieve the next set of results.

$sel:pipelineId:DescribeObjects', describeObjects_pipelineId - The ID of the pipeline that contains the object definitions.

$sel:objectIds:DescribeObjects', describeObjects_objectIds - The IDs of the pipeline objects that contain the definitions to be described. You can pass as many as 25 identifiers in a single call to DescribeObjects.

Request Lenses

describeObjects_evaluateExpressions :: Lens' DescribeObjects (Maybe Bool)

Indicates whether any expressions in the object should be evaluated when the object descriptions are returned.

describeObjects_marker :: Lens' DescribeObjects (Maybe Text)

The starting point for the results to be returned. For the first call, this value should be empty. As long as there are more results, continue to call DescribeObjects with the marker value from the previous call to retrieve the next set of results.

describeObjects_pipelineId :: Lens' DescribeObjects Text

The ID of the pipeline that contains the object definitions.

describeObjects_objectIds :: Lens' DescribeObjects [Text]

The IDs of the pipeline objects that contain the definitions to be described. You can pass as many as 25 identifiers in a single call to DescribeObjects.

Destructuring the Response

data DescribeObjectsResponse

Contains the output of DescribeObjects.

See: newDescribeObjectsResponse smart constructor.

Constructors

DescribeObjectsResponse'

Fields

hasMoreResults :: Maybe Bool

Indicates whether there are more results to return.

marker :: Maybe Text

The starting point for the next page of results. To view the next page of results, call DescribeObjects again with this marker value. If the value is null, there are no more results.

httpStatus :: Int

The response's http status code.

pipelineObjects :: [PipelineObject]

An array of object definitions.

newDescribeObjectsResponse

Arguments

:: Int
$sel:httpStatus:DescribeObjectsResponse'

-> DescribeObjectsResponse

Create a value of DescribeObjectsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hasMoreResults:DescribeObjectsResponse', describeObjectsResponse_hasMoreResults - Indicates whether there are more results to return.

$sel:marker:DescribeObjects', describeObjectsResponse_marker - The starting point for the next page of results. To view the next page of results, call DescribeObjects again with this marker value. If the value is null, there are no more results.

$sel:httpStatus:DescribeObjectsResponse', describeObjectsResponse_httpStatus - The response's http status code.

$sel:pipelineObjects:DescribeObjectsResponse', describeObjectsResponse_pipelineObjects - An array of object definitions.

Response Lenses

describeObjectsResponse_hasMoreResults :: Lens' DescribeObjectsResponse (Maybe Bool)

Indicates whether there are more results to return.

describeObjectsResponse_marker :: Lens' DescribeObjectsResponse (Maybe Text)

The starting point for the next page of results. To view the next page of results, call DescribeObjects again with this marker value. If the value is null, there are no more results.

describeObjectsResponse_httpStatus :: Lens' DescribeObjectsResponse Int

The response's http status code.

describeObjectsResponse_pipelineObjects :: Lens' DescribeObjectsResponse [PipelineObject]

An array of object definitions.