Network.AWS.DataPipeline.Lens

Description

Operations

RemoveTags

removeTags_pipelineId :: Lens' RemoveTags Text

The ID of the pipeline.

removeTags_tagKeys :: Lens' RemoveTags [Text]

The keys of the tags to remove.

removeTagsResponse_httpStatus :: Lens' RemoveTagsResponse Int

The response's http status code.

QueryObjects

queryObjects_query :: Lens' QueryObjects (Maybe Query)

The query that defines the objects to be returned. The Query object can contain a maximum of ten selectors. The conditions in the query are limited to top-level String fields in the object. These filters can be applied to components, instances, and attempts.

queryObjects_limit :: Lens' QueryObjects (Maybe Int)

The maximum number of object names that QueryObjects will return in a single call. The default value is 100.

queryObjects_marker :: Lens' QueryObjects (Maybe Text)

The starting point for the results to be returned. For the first call, this value should be empty. As long as there are more results, continue to call QueryObjects with the marker value from the previous call to retrieve the next set of results.

queryObjects_pipelineId :: Lens' QueryObjects Text

The ID of the pipeline.

queryObjects_sphere :: Lens' QueryObjects Text

Indicates whether the query applies to components or instances. The possible values are: COMPONENT, INSTANCE, and ATTEMPT.

queryObjectsResponse_ids :: Lens' QueryObjectsResponse (Maybe [Text])

The identifiers that match the query selectors.

queryObjectsResponse_hasMoreResults :: Lens' QueryObjectsResponse (Maybe Bool)

Indicates whether there are more results that can be obtained by a subsequent call.

queryObjectsResponse_marker :: Lens' QueryObjectsResponse (Maybe Text)

The starting point for the next page of results. To view the next page of results, call QueryObjects again with this marker value. If the value is null, there are no more results.

queryObjectsResponse_httpStatus :: Lens' QueryObjectsResponse Int

The response's http status code.

EvaluateExpression

evaluateExpression_pipelineId :: Lens' EvaluateExpression Text

The ID of the pipeline.

evaluateExpression_objectId :: Lens' EvaluateExpression Text

The ID of the object.

evaluateExpression_expression :: Lens' EvaluateExpression Text

The expression to evaluate.

evaluateExpressionResponse_httpStatus :: Lens' EvaluateExpressionResponse Int

The response's http status code.

evaluateExpressionResponse_evaluatedExpression :: Lens' EvaluateExpressionResponse Text

The evaluated expression.

ReportTaskProgress

reportTaskProgress_fields :: Lens' ReportTaskProgress (Maybe [Field])

Key-value pairs that define the properties of the ReportTaskProgressInput object.

reportTaskProgress_taskId :: Lens' ReportTaskProgress Text

The ID of the task assigned to the task runner. This value is provided in the response for PollForTask.

reportTaskProgressResponse_httpStatus :: Lens' ReportTaskProgressResponse Int

The response's http status code.

reportTaskProgressResponse_canceled :: Lens' ReportTaskProgressResponse Bool

If true, the calling task runner should cancel processing of the task. The task runner does not need to call SetTaskStatus for canceled tasks.

CreatePipeline

createPipeline_tags :: Lens' CreatePipeline (Maybe [Tag])

A list of tags to associate with the pipeline at creation. Tags let you control access to pipelines. For more information, see Controlling User Access to Pipelines in the AWS Data Pipeline Developer Guide.

createPipeline_description :: Lens' CreatePipeline (Maybe Text)

The description for the pipeline.

createPipeline_name :: Lens' CreatePipeline Text

The name for the pipeline. You can use the same name for multiple pipelines associated with your AWS account, because AWS Data Pipeline assigns each pipeline a unique pipeline identifier.

createPipeline_uniqueId :: Lens' CreatePipeline Text

A unique identifier. This identifier is not the same as the pipeline identifier assigned by AWS Data Pipeline. You are responsible for defining the format and ensuring the uniqueness of this identifier. You use this parameter to ensure idempotency during repeated calls to CreatePipeline. For example, if the first call to CreatePipeline does not succeed, you can pass in the same unique identifier and pipeline name combination on a subsequent call to CreatePipeline. CreatePipeline ensures that if a pipeline already exists with the same name and unique identifier, a new pipeline is not created. Instead, you'll receive the pipeline identifier from the previous attempt. The uniqueness of the name and unique identifier combination is scoped to the AWS account or IAM user credentials.

createPipelineResponse_httpStatus :: Lens' CreatePipelineResponse Int

The response's http status code.

createPipelineResponse_pipelineId :: Lens' CreatePipelineResponse Text

The ID that AWS Data Pipeline assigns the newly created pipeline. For example, df-06372391ZG65EXAMPLE.

DeletePipeline

deletePipeline_pipelineId :: Lens' DeletePipeline Text

The ID of the pipeline.

SetTaskStatus

setTaskStatus_errorStackTrace :: Lens' SetTaskStatus (Maybe Text)

If an error occurred during the task, this value specifies the stack trace associated with the error. This value is set on the physical attempt object. It is used to display error information to the user. The web service does not parse this value.

setTaskStatus_errorMessage :: Lens' SetTaskStatus (Maybe Text)

If an error occurred during the task, this value specifies a text description of the error. This value is set on the physical attempt object. It is used to display error information to the user. The web service does not parse this value.

setTaskStatus_errorId :: Lens' SetTaskStatus (Maybe Text)

If an error occurred during the task, this value specifies the error code. This value is set on the physical attempt object. It is used to display error information to the user. It should not start with string "Service_" which is reserved by the system.

setTaskStatus_taskId :: Lens' SetTaskStatus Text

The ID of the task assigned to the task runner. This value is provided in the response for PollForTask.

setTaskStatus_taskStatus :: Lens' SetTaskStatus TaskStatus

If FINISHED, the task successfully completed. If FAILED, the task ended unsuccessfully. Preconditions use false.

setTaskStatusResponse_httpStatus :: Lens' SetTaskStatusResponse Int

The response's http status code.

ReportTaskRunnerHeartbeat

reportTaskRunnerHeartbeat_hostname :: Lens' ReportTaskRunnerHeartbeat (Maybe Text)

The public DNS name of the task runner.

reportTaskRunnerHeartbeat_workerGroup :: Lens' ReportTaskRunnerHeartbeat (Maybe Text)

The type of task the task runner is configured to accept and process. The worker group is set as a field on objects in the pipeline when they are created. You can only specify a single value for workerGroup. There are no wildcard values permitted in workerGroup; the string must be an exact, case-sensitive, match.

reportTaskRunnerHeartbeat_taskrunnerId :: Lens' ReportTaskRunnerHeartbeat Text

The ID of the task runner. This value should be unique across your AWS account. In the case of AWS Data Pipeline Task Runner launched on a resource managed by AWS Data Pipeline, the web service provides a unique identifier when it launches the application. If you have written a custom task runner, you should assign a unique identifier for the task runner.

reportTaskRunnerHeartbeatResponse_httpStatus :: Lens' ReportTaskRunnerHeartbeatResponse Int

The response's http status code.

reportTaskRunnerHeartbeatResponse_terminate :: Lens' ReportTaskRunnerHeartbeatResponse Bool

Indicates whether the calling task runner should terminate.

DescribeObjects

describeObjects_evaluateExpressions :: Lens' DescribeObjects (Maybe Bool)

Indicates whether any expressions in the object should be evaluated when the object descriptions are returned.

describeObjects_marker :: Lens' DescribeObjects (Maybe Text)

The starting point for the results to be returned. For the first call, this value should be empty. As long as there are more results, continue to call DescribeObjects with the marker value from the previous call to retrieve the next set of results.

describeObjects_pipelineId :: Lens' DescribeObjects Text

The ID of the pipeline that contains the object definitions.

describeObjects_objectIds :: Lens' DescribeObjects [Text]

The IDs of the pipeline objects that contain the definitions to be described. You can pass as many as 25 identifiers in a single call to DescribeObjects.

describeObjectsResponse_hasMoreResults :: Lens' DescribeObjectsResponse (Maybe Bool)

Indicates whether there are more results to return.

describeObjectsResponse_marker :: Lens' DescribeObjectsResponse (Maybe Text)

The starting point for the next page of results. To view the next page of results, call DescribeObjects again with this marker value. If the value is null, there are no more results.

describeObjectsResponse_httpStatus :: Lens' DescribeObjectsResponse Int

The response's http status code.

describeObjectsResponse_pipelineObjects :: Lens' DescribeObjectsResponse [PipelineObject]

An array of object definitions.

AddTags

addTags_pipelineId :: Lens' AddTags Text

The ID of the pipeline.

addTags_tags :: Lens' AddTags [Tag]

The tags to add, as key/value pairs.

addTagsResponse_httpStatus :: Lens' AddTagsResponse Int

The response's http status code.

DescribePipelines

describePipelines_pipelineIds :: Lens' DescribePipelines [Text]

The IDs of the pipelines to describe. You can pass as many as 25 identifiers in a single call. To obtain pipeline IDs, call ListPipelines.

describePipelinesResponse_httpStatus :: Lens' DescribePipelinesResponse Int

The response's http status code.

describePipelinesResponse_pipelineDescriptionList :: Lens' DescribePipelinesResponse [PipelineDescription]

An array of descriptions for the specified pipelines.

PollForTask

pollForTask_hostname :: Lens' PollForTask (Maybe Text)

The public DNS name of the calling task runner.

pollForTask_instanceIdentity :: Lens' PollForTask (Maybe InstanceIdentity)

Identity information for the EC2 instance that is hosting the task runner. You can get this value from the instance using http://169.254.169.254/latest/meta-data/instance-id. For more information, see Instance Metadata in the Amazon Elastic Compute Cloud User Guide. Passing in this value proves that your task runner is running on an EC2 instance, and ensures the proper AWS Data Pipeline service charges are applied to your pipeline.

pollForTask_workerGroup :: Lens' PollForTask Text

The type of task the task runner is configured to accept and process. The worker group is set as a field on objects in the pipeline when they are created. You can only specify a single value for workerGroup in the call to PollForTask. There are no wildcard values permitted in workerGroup; the string must be an exact, case-sensitive, match.

pollForTaskResponse_taskObject :: Lens' PollForTaskResponse (Maybe TaskObject)

The information needed to complete the task that is being assigned to the task runner. One of the fields returned in this object is taskId, which contains an identifier for the task being assigned. The calling task runner uses taskId in subsequent calls to ReportTaskProgress and SetTaskStatus.

pollForTaskResponse_httpStatus :: Lens' PollForTaskResponse Int

The response's http status code.

ValidatePipelineDefinition

validatePipelineDefinition_parameterValues :: Lens' ValidatePipelineDefinition (Maybe [ParameterValue])

The parameter values used with the pipeline.

validatePipelineDefinition_parameterObjects :: Lens' ValidatePipelineDefinition (Maybe [ParameterObject])

The parameter objects used with the pipeline.

validatePipelineDefinition_pipelineId :: Lens' ValidatePipelineDefinition Text

The ID of the pipeline.

validatePipelineDefinition_pipelineObjects :: Lens' ValidatePipelineDefinition [PipelineObject]

The objects that define the pipeline changes to validate against the pipeline.

validatePipelineDefinitionResponse_validationErrors :: Lens' ValidatePipelineDefinitionResponse (Maybe [ValidationError])

Any validation errors that were found.

validatePipelineDefinitionResponse_validationWarnings :: Lens' ValidatePipelineDefinitionResponse (Maybe [ValidationWarning])

Any validation warnings that were found.

validatePipelineDefinitionResponse_httpStatus :: Lens' ValidatePipelineDefinitionResponse Int

The response's http status code.

validatePipelineDefinitionResponse_errored :: Lens' ValidatePipelineDefinitionResponse Bool

Indicates whether there were validation errors.

GetPipelineDefinition

getPipelineDefinition_version :: Lens' GetPipelineDefinition (Maybe Text)

The version of the pipeline definition to retrieve. Set this parameter to latest (default) to use the last definition saved to the pipeline or active to use the last definition that was activated.

getPipelineDefinition_pipelineId :: Lens' GetPipelineDefinition Text

The ID of the pipeline.

getPipelineDefinitionResponse_parameterValues :: Lens' GetPipelineDefinitionResponse (Maybe [ParameterValue])

The parameter values used in the pipeline definition.

getPipelineDefinitionResponse_parameterObjects :: Lens' GetPipelineDefinitionResponse (Maybe [ParameterObject])

The parameter objects used in the pipeline definition.

getPipelineDefinitionResponse_pipelineObjects :: Lens' GetPipelineDefinitionResponse (Maybe [PipelineObject])

The objects defined in the pipeline.

getPipelineDefinitionResponse_httpStatus :: Lens' GetPipelineDefinitionResponse Int

The response's http status code.

PutPipelineDefinition

putPipelineDefinition_parameterValues :: Lens' PutPipelineDefinition (Maybe [ParameterValue])

The parameter values used with the pipeline.

putPipelineDefinition_parameterObjects :: Lens' PutPipelineDefinition (Maybe [ParameterObject])

The parameter objects used with the pipeline.

putPipelineDefinition_pipelineId :: Lens' PutPipelineDefinition Text

The ID of the pipeline.

putPipelineDefinition_pipelineObjects :: Lens' PutPipelineDefinition [PipelineObject]

The objects that define the pipeline. These objects overwrite the existing pipeline definition.

putPipelineDefinitionResponse_validationErrors :: Lens' PutPipelineDefinitionResponse (Maybe [ValidationError])

The validation errors that are associated with the objects defined in pipelineObjects.

putPipelineDefinitionResponse_validationWarnings :: Lens' PutPipelineDefinitionResponse (Maybe [ValidationWarning])

The validation warnings that are associated with the objects defined in pipelineObjects.

putPipelineDefinitionResponse_httpStatus :: Lens' PutPipelineDefinitionResponse Int

The response's http status code.

putPipelineDefinitionResponse_errored :: Lens' PutPipelineDefinitionResponse Bool

Indicates whether there were validation errors, and the pipeline definition is stored but cannot be activated until you correct the pipeline and call PutPipelineDefinition to commit the corrected pipeline.

SetStatus

setStatus_pipelineId :: Lens' SetStatus Text

The ID of the pipeline that contains the objects.

setStatus_objectIds :: Lens' SetStatus [Text]

The IDs of the objects. The corresponding objects can be either physical or components, but not a mix of both types.

setStatus_status :: Lens' SetStatus Text

The status to be set on all the objects specified in objectIds. For components, use PAUSE or RESUME. For instances, use TRY_CANCEL, RERUN, or MARK_FINISHED.

ListPipelines

listPipelines_marker :: Lens' ListPipelines (Maybe Text)

The starting point for the results to be returned. For the first call, this value should be empty. As long as there are more results, continue to call ListPipelines with the marker value from the previous call to retrieve the next set of results.

listPipelinesResponse_hasMoreResults :: Lens' ListPipelinesResponse (Maybe Bool)

Indicates whether there are more results that can be obtained by a subsequent call.

listPipelinesResponse_marker :: Lens' ListPipelinesResponse (Maybe Text)

The starting point for the next page of results. To view the next page of results, call ListPipelinesOutput again with this marker value. If the value is null, there are no more results.

listPipelinesResponse_httpStatus :: Lens' ListPipelinesResponse Int

The response's http status code.

listPipelinesResponse_pipelineIdList :: Lens' ListPipelinesResponse [PipelineIdName]

The pipeline identifiers. If you require additional information about the pipelines, you can use these identifiers to call DescribePipelines and GetPipelineDefinition.

ActivatePipeline

activatePipeline_parameterValues :: Lens' ActivatePipeline (Maybe [ParameterValue])

A list of parameter values to pass to the pipeline at activation.

activatePipeline_startTimestamp :: Lens' ActivatePipeline (Maybe UTCTime)

The date and time to resume the pipeline. By default, the pipeline resumes from the last completed execution.

activatePipeline_pipelineId :: Lens' ActivatePipeline Text

The ID of the pipeline.

activatePipelineResponse_httpStatus :: Lens' ActivatePipelineResponse Int

The response's http status code.

DeactivatePipeline

deactivatePipeline_cancelActive :: Lens' DeactivatePipeline (Maybe Bool)

Indicates whether to cancel any running objects. The default is true, which sets the state of any running objects to CANCELED. If this value is false, the pipeline is deactivated after all running objects finish.

deactivatePipeline_pipelineId :: Lens' DeactivatePipeline Text

The ID of the pipeline.

deactivatePipelineResponse_httpStatus :: Lens' DeactivatePipelineResponse Int

The response's http status code.

Types

Field

field_stringValue :: Lens' Field (Maybe Text)

The field value, expressed as a String.

field_refValue :: Lens' Field (Maybe Text)

The field value, expressed as the identifier of another object.

field_key :: Lens' Field Text

The field identifier.

InstanceIdentity

instanceIdentity_document :: Lens' InstanceIdentity (Maybe Text)

A description of an EC2 instance that is generated when the instance is launched and exposed to the instance via the instance metadata service in the form of a JSON representation of an object.

instanceIdentity_signature :: Lens' InstanceIdentity (Maybe Text)

A signature which can be used to verify the accuracy and authenticity of the information provided in the instance identity document.

Operator

operator_values :: Lens' Operator (Maybe [Text])

The value that the actual field value will be compared with.

operator_type :: Lens' Operator (Maybe OperatorType)

The logical operation to be performed: equal (EQ), equal reference (REF_EQ), less than or equal (LE), greater than or equal (GE), or between (BETWEEN). Equal reference (REF_EQ) can be used only with reference fields. The other comparison types can be used only with String fields. The comparison types you can use apply only to certain object fields, as detailed below.

The comparison operators EQ and REF_EQ act on the following fields:

name

@sphere

parent

@componentParent

@instanceParent

@status

@scheduledStartTime

@scheduledEndTime

@actualStartTime

@actualEndTime

The comparison operators GE, LE, and BETWEEN act on the following fields:

@scheduledStartTime

@scheduledEndTime

@actualStartTime

@actualEndTime

Note that fields beginning with the at sign (@) are read-only and set by the web service. When you name fields, you should choose names containing only alpha-numeric values, as symbols may be reserved by AWS Data Pipeline. User-defined fields that you add to a pipeline should prefix their name with the string "my".

ParameterAttribute

parameterAttribute_key :: Lens' ParameterAttribute Text

The field identifier.

parameterAttribute_stringValue :: Lens' ParameterAttribute Text

The field value, expressed as a String.

ParameterObject

parameterObject_id :: Lens' ParameterObject Text

The ID of the parameter object.

parameterObject_attributes :: Lens' ParameterObject [ParameterAttribute]

The attributes of the parameter object.

ParameterValue

parameterValue_id :: Lens' ParameterValue Text

The ID of the parameter value.

parameterValue_stringValue :: Lens' ParameterValue Text

The field value, expressed as a String.

PipelineDescription

pipelineDescription_tags :: Lens' PipelineDescription (Maybe [Tag])

A list of tags to associated with a pipeline. Tags let you control access to pipelines. For more information, see Controlling User Access to Pipelines in the AWS Data Pipeline Developer Guide.

pipelineDescription_description :: Lens' PipelineDescription (Maybe Text)

Description of the pipeline.

pipelineDescription_pipelineId :: Lens' PipelineDescription Text

The pipeline identifier that was assigned by AWS Data Pipeline. This is a string of the form df-297EG78HU43EEXAMPLE.

pipelineDescription_name :: Lens' PipelineDescription Text

The name of the pipeline.

pipelineDescription_fields :: Lens' PipelineDescription [Field]

A list of read-only fields that contain metadata about the pipeline: @userId, @accountId, and @pipelineState.

PipelineIdName

pipelineIdName_id :: Lens' PipelineIdName (Maybe Text)

The ID of the pipeline that was assigned by AWS Data Pipeline. This is a string of the form df-297EG78HU43EEXAMPLE.

pipelineIdName_name :: Lens' PipelineIdName (Maybe Text)

The name of the pipeline.

PipelineObject

pipelineObject_id :: Lens' PipelineObject Text

The ID of the object.

pipelineObject_name :: Lens' PipelineObject Text

The name of the object.

pipelineObject_fields :: Lens' PipelineObject [Field]

Key-value pairs that define the properties of the object.

Query

query_selectors :: Lens' Query (Maybe [Selector])

List of selectors that define the query. An object must satisfy all of the selectors to match the query.

Selector

selector_operator :: Lens' Selector (Maybe Operator)

Undocumented member.

selector_fieldName :: Lens' Selector (Maybe Text)

The name of the field that the operator will be applied to. The field name is the "key" portion of the field definition in the pipeline definition syntax that is used by the AWS Data Pipeline API. If the field is not set on the object, the condition fails.

Tag

tag_key :: Lens' Tag Text

The key name of a tag defined by a user. For more information, see Controlling User Access to Pipelines in the AWS Data Pipeline Developer Guide.

tag_value :: Lens' Tag Text

The optional value portion of a tag defined by a user. For more information, see Controlling User Access to Pipelines in the AWS Data Pipeline Developer Guide.

TaskObject

taskObject_pipelineId :: Lens' TaskObject (Maybe Text)

The ID of the pipeline that provided the task.

taskObject_objects :: Lens' TaskObject (Maybe (HashMap Text PipelineObject))

Connection information for the location where the task runner will publish the output of the task.

taskObject_taskId :: Lens' TaskObject (Maybe Text)

An internal identifier for the task. This ID is passed to the SetTaskStatus and ReportTaskProgress actions.

taskObject_attemptId :: Lens' TaskObject (Maybe Text)

The ID of the pipeline task attempt object. AWS Data Pipeline uses this value to track how many times a task is attempted.

ValidationError

validationError_id :: Lens' ValidationError (Maybe Text)

The identifier of the object that contains the validation error.

validationError_errors :: Lens' ValidationError (Maybe [Text])

A description of the validation error.

ValidationWarning

validationWarning_warnings :: Lens' ValidationWarning (Maybe [Text])

A description of the validation warning.

validationWarning_id :: Lens' ValidationWarning (Maybe Text)

The identifier of the object that contains the validation warning.