Network.AWS.DataPipeline.ListPipelines

Description

Lists the pipeline identifiers for all active pipelines that you have permission to access.

This operation returns paginated results.

Creating a Request

data ListPipelines

Contains the parameters for ListPipelines.

See: newListPipelines smart constructor.

Constructors

ListPipelines'

Fields

marker :: Maybe Text

The starting point for the results to be returned. For the first call, this value should be empty. As long as there are more results, continue to call ListPipelines with the marker value from the previous call to retrieve the next set of results.

newListPipelines :: ListPipelines

Create a value of ListPipelines with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:marker:ListPipelines', listPipelines_marker - The starting point for the results to be returned. For the first call, this value should be empty. As long as there are more results, continue to call ListPipelines with the marker value from the previous call to retrieve the next set of results.

Request Lenses

listPipelines_marker :: Lens' ListPipelines (Maybe Text)

The starting point for the results to be returned. For the first call, this value should be empty. As long as there are more results, continue to call ListPipelines with the marker value from the previous call to retrieve the next set of results.

Destructuring the Response

data ListPipelinesResponse

Contains the output of ListPipelines.

See: newListPipelinesResponse smart constructor.

Constructors

ListPipelinesResponse'

Fields

hasMoreResults :: Maybe Bool

Indicates whether there are more results that can be obtained by a subsequent call.

marker :: Maybe Text

The starting point for the next page of results. To view the next page of results, call ListPipelinesOutput again with this marker value. If the value is null, there are no more results.

httpStatus :: Int

The response's http status code.

pipelineIdList :: [PipelineIdName]

The pipeline identifiers. If you require additional information about the pipelines, you can use these identifiers to call DescribePipelines and GetPipelineDefinition.

newListPipelinesResponse

Arguments

:: Int
$sel:httpStatus:ListPipelinesResponse'

-> ListPipelinesResponse

Create a value of ListPipelinesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hasMoreResults:ListPipelinesResponse', listPipelinesResponse_hasMoreResults - Indicates whether there are more results that can be obtained by a subsequent call.

$sel:marker:ListPipelines', listPipelinesResponse_marker - The starting point for the next page of results. To view the next page of results, call ListPipelinesOutput again with this marker value. If the value is null, there are no more results.

$sel:httpStatus:ListPipelinesResponse', listPipelinesResponse_httpStatus - The response's http status code.

$sel:pipelineIdList:ListPipelinesResponse', listPipelinesResponse_pipelineIdList - The pipeline identifiers. If you require additional information about the pipelines, you can use these identifiers to call DescribePipelines and GetPipelineDefinition.

Response Lenses

listPipelinesResponse_hasMoreResults :: Lens' ListPipelinesResponse (Maybe Bool)

Indicates whether there are more results that can be obtained by a subsequent call.

listPipelinesResponse_marker :: Lens' ListPipelinesResponse (Maybe Text)

The starting point for the next page of results. To view the next page of results, call ListPipelinesOutput again with this marker value. If the value is null, there are no more results.

listPipelinesResponse_httpStatus :: Lens' ListPipelinesResponse Int

The response's http status code.

listPipelinesResponse_pipelineIdList :: Lens' ListPipelinesResponse [PipelineIdName]

The pipeline identifiers. If you require additional information about the pipelines, you can use these identifiers to call DescribePipelines and GetPipelineDefinition.