Network.AWS.DataPipeline.PollForTask

Description

Task runners call PollForTask to receive a task to perform from AWS Data Pipeline. The task runner specifies which tasks it can perform by setting a value for the workerGroup parameter. The task returned can come from any of the pipelines that match the workerGroup value passed in by the task runner and that was launched using the IAM user credentials specified by the task runner.

If tasks are ready in the work queue, PollForTask returns a response immediately. If no tasks are available in the queue, PollForTask uses long-polling and holds on to a poll connection for up to a 90 seconds, during which time the first newly scheduled task is handed to the task runner. To accomodate this, set the socket timeout in your task runner to 90 seconds. The task runner should not call PollForTask again on the same workerGroup until it receives a response, and this can take up to 90 seconds.

Creating a Request

data PollForTask

Contains the parameters for PollForTask.

See: newPollForTask smart constructor.

Constructors

PollForTask'

Fields

hostname :: Maybe Text

The public DNS name of the calling task runner.

instanceIdentity :: Maybe InstanceIdentity

Identity information for the EC2 instance that is hosting the task runner. You can get this value from the instance using http://169.254.169.254/latest/meta-data/instance-id. For more information, see Instance Metadata in the Amazon Elastic Compute Cloud User Guide. Passing in this value proves that your task runner is running on an EC2 instance, and ensures the proper AWS Data Pipeline service charges are applied to your pipeline.

workerGroup :: Text

The type of task the task runner is configured to accept and process. The worker group is set as a field on objects in the pipeline when they are created. You can only specify a single value for workerGroup in the call to PollForTask. There are no wildcard values permitted in workerGroup; the string must be an exact, case-sensitive, match.

newPollForTask

Arguments

:: Text
$sel:workerGroup:PollForTask'

-> PollForTask

Create a value of PollForTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hostname:PollForTask', pollForTask_hostname - The public DNS name of the calling task runner.

$sel:instanceIdentity:PollForTask', pollForTask_instanceIdentity - Identity information for the EC2 instance that is hosting the task runner. You can get this value from the instance using http://169.254.169.254/latest/meta-data/instance-id. For more information, see Instance Metadata in the Amazon Elastic Compute Cloud User Guide. Passing in this value proves that your task runner is running on an EC2 instance, and ensures the proper AWS Data Pipeline service charges are applied to your pipeline.

$sel:workerGroup:PollForTask', pollForTask_workerGroup - The type of task the task runner is configured to accept and process. The worker group is set as a field on objects in the pipeline when they are created. You can only specify a single value for workerGroup in the call to PollForTask. There are no wildcard values permitted in workerGroup; the string must be an exact, case-sensitive, match.

Request Lenses

pollForTask_hostname :: Lens' PollForTask (Maybe Text)

The public DNS name of the calling task runner.

pollForTask_instanceIdentity :: Lens' PollForTask (Maybe InstanceIdentity)

Identity information for the EC2 instance that is hosting the task runner. You can get this value from the instance using http://169.254.169.254/latest/meta-data/instance-id. For more information, see Instance Metadata in the Amazon Elastic Compute Cloud User Guide. Passing in this value proves that your task runner is running on an EC2 instance, and ensures the proper AWS Data Pipeline service charges are applied to your pipeline.

pollForTask_workerGroup :: Lens' PollForTask Text

The type of task the task runner is configured to accept and process. The worker group is set as a field on objects in the pipeline when they are created. You can only specify a single value for workerGroup in the call to PollForTask. There are no wildcard values permitted in workerGroup; the string must be an exact, case-sensitive, match.

Destructuring the Response

data PollForTaskResponse

Contains the output of PollForTask.

See: newPollForTaskResponse smart constructor.

Constructors

PollForTaskResponse'

Fields

taskObject :: Maybe TaskObject

The information needed to complete the task that is being assigned to the task runner. One of the fields returned in this object is taskId, which contains an identifier for the task being assigned. The calling task runner uses taskId in subsequent calls to ReportTaskProgress and SetTaskStatus.

httpStatus :: Int

The response's http status code.

newPollForTaskResponse

Arguments

:: Int
$sel:httpStatus:PollForTaskResponse'

-> PollForTaskResponse

Create a value of PollForTaskResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:taskObject:PollForTaskResponse', pollForTaskResponse_taskObject - The information needed to complete the task that is being assigned to the task runner. One of the fields returned in this object is taskId, which contains an identifier for the task being assigned. The calling task runner uses taskId in subsequent calls to ReportTaskProgress and SetTaskStatus.

$sel:httpStatus:PollForTaskResponse', pollForTaskResponse_httpStatus - The response's http status code.

Response Lenses

pollForTaskResponse_taskObject :: Lens' PollForTaskResponse (Maybe TaskObject)

The information needed to complete the task that is being assigned to the task runner. One of the fields returned in this object is taskId, which contains an identifier for the task being assigned. The calling task runner uses taskId in subsequent calls to ReportTaskProgress and SetTaskStatus.

pollForTaskResponse_httpStatus :: Lens' PollForTaskResponse Int

The response's http status code.