Network.AWS.DataPipeline.PutPipelineDefinition

Description

Adds tasks, schedules, and preconditions to the specified pipeline. You can use PutPipelineDefinition to populate a new pipeline.

PutPipelineDefinition also validates the configuration as it adds it to the pipeline. Changes to the pipeline are saved unless one of the following three validation errors exists in the pipeline.

An object is missing a name or identifier field.

A string or reference field is empty.

The number of objects in the pipeline exceeds the maximum allowed objects.

The pipeline is in a FINISHED state.

Pipeline object definitions are passed to the PutPipelineDefinition action and returned by the GetPipelineDefinition action.

Creating a Request

data PutPipelineDefinition

Contains the parameters for PutPipelineDefinition.

See: newPutPipelineDefinition smart constructor.

Constructors

PutPipelineDefinition'

Fields

parameterObjects :: Maybe [ParameterObject]

The parameter objects used with the pipeline.

parameterValues :: Maybe [ParameterValue]

The parameter values used with the pipeline.

pipelineId :: Text

The ID of the pipeline.

pipelineObjects :: [PipelineObject]

The objects that define the pipeline. These objects overwrite the existing pipeline definition.

newPutPipelineDefinition

Arguments

:: Text
$sel:pipelineId:PutPipelineDefinition'

-> PutPipelineDefinition

Create a value of PutPipelineDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parameterObjects:PutPipelineDefinition', putPipelineDefinition_parameterObjects - The parameter objects used with the pipeline.

$sel:parameterValues:PutPipelineDefinition', putPipelineDefinition_parameterValues - The parameter values used with the pipeline.

$sel:pipelineId:PutPipelineDefinition', putPipelineDefinition_pipelineId - The ID of the pipeline.

$sel:pipelineObjects:PutPipelineDefinition', putPipelineDefinition_pipelineObjects - The objects that define the pipeline. These objects overwrite the existing pipeline definition.

Request Lenses

putPipelineDefinition_parameterObjects :: Lens' PutPipelineDefinition (Maybe [ParameterObject])

The parameter objects used with the pipeline.

putPipelineDefinition_parameterValues :: Lens' PutPipelineDefinition (Maybe [ParameterValue])

The parameter values used with the pipeline.

putPipelineDefinition_pipelineId :: Lens' PutPipelineDefinition Text

The ID of the pipeline.

putPipelineDefinition_pipelineObjects :: Lens' PutPipelineDefinition [PipelineObject]

The objects that define the pipeline. These objects overwrite the existing pipeline definition.

Destructuring the Response

data PutPipelineDefinitionResponse

Contains the output of PutPipelineDefinition.

See: newPutPipelineDefinitionResponse smart constructor.

Constructors

PutPipelineDefinitionResponse'

Fields

validationErrors :: Maybe [ValidationError]

The validation errors that are associated with the objects defined in pipelineObjects.

validationWarnings :: Maybe [ValidationWarning]

The validation warnings that are associated with the objects defined in pipelineObjects.

httpStatus :: Int

The response's http status code.

errored :: Bool

Indicates whether there were validation errors, and the pipeline definition is stored but cannot be activated until you correct the pipeline and call PutPipelineDefinition to commit the corrected pipeline.

newPutPipelineDefinitionResponse

Arguments

:: Int
$sel:httpStatus:PutPipelineDefinitionResponse'

-> Bool
$sel:errored:PutPipelineDefinitionResponse'

-> PutPipelineDefinitionResponse

Create a value of PutPipelineDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:validationErrors:PutPipelineDefinitionResponse', putPipelineDefinitionResponse_validationErrors - The validation errors that are associated with the objects defined in pipelineObjects.

$sel:validationWarnings:PutPipelineDefinitionResponse', putPipelineDefinitionResponse_validationWarnings - The validation warnings that are associated with the objects defined in pipelineObjects.

$sel:httpStatus:PutPipelineDefinitionResponse', putPipelineDefinitionResponse_httpStatus - The response's http status code.

$sel:errored:PutPipelineDefinitionResponse', putPipelineDefinitionResponse_errored - Indicates whether there were validation errors, and the pipeline definition is stored but cannot be activated until you correct the pipeline and call PutPipelineDefinition to commit the corrected pipeline.

Response Lenses

putPipelineDefinitionResponse_validationErrors :: Lens' PutPipelineDefinitionResponse (Maybe [ValidationError])

The validation errors that are associated with the objects defined in pipelineObjects.

putPipelineDefinitionResponse_validationWarnings :: Lens' PutPipelineDefinitionResponse (Maybe [ValidationWarning])

The validation warnings that are associated with the objects defined in pipelineObjects.

putPipelineDefinitionResponse_httpStatus :: Lens' PutPipelineDefinitionResponse Int

The response's http status code.

putPipelineDefinitionResponse_errored :: Lens' PutPipelineDefinitionResponse Bool

Indicates whether there were validation errors, and the pipeline definition is stored but cannot be activated until you correct the pipeline and call PutPipelineDefinition to commit the corrected pipeline.