Network.AWS.DataPipeline.QueryObjects

Description

Queries the specified pipeline for the names of objects that match the specified set of conditions.

This operation returns paginated results.

Creating a Request

data QueryObjects

Contains the parameters for QueryObjects.

See: newQueryObjects smart constructor.

Constructors

QueryObjects'

Fields

query :: Maybe Query

The query that defines the objects to be returned. The Query object can contain a maximum of ten selectors. The conditions in the query are limited to top-level String fields in the object. These filters can be applied to components, instances, and attempts.

limit :: Maybe Int

The maximum number of object names that QueryObjects will return in a single call. The default value is 100.

marker :: Maybe Text

The starting point for the results to be returned. For the first call, this value should be empty. As long as there are more results, continue to call QueryObjects with the marker value from the previous call to retrieve the next set of results.

pipelineId :: Text

The ID of the pipeline.

sphere :: Text

Indicates whether the query applies to components or instances. The possible values are: COMPONENT, INSTANCE, and ATTEMPT.

newQueryObjects

Arguments

:: Text
$sel:pipelineId:QueryObjects'

-> Text
$sel:sphere:QueryObjects'

-> QueryObjects

Create a value of QueryObjects with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:query:QueryObjects', queryObjects_query - The query that defines the objects to be returned. The Query object can contain a maximum of ten selectors. The conditions in the query are limited to top-level String fields in the object. These filters can be applied to components, instances, and attempts.

$sel:limit:QueryObjects', queryObjects_limit - The maximum number of object names that QueryObjects will return in a single call. The default value is 100.

$sel:marker:QueryObjects', queryObjects_marker - The starting point for the results to be returned. For the first call, this value should be empty. As long as there are more results, continue to call QueryObjects with the marker value from the previous call to retrieve the next set of results.

$sel:pipelineId:QueryObjects', queryObjects_pipelineId - The ID of the pipeline.

$sel:sphere:QueryObjects', queryObjects_sphere - Indicates whether the query applies to components or instances. The possible values are: COMPONENT, INSTANCE, and ATTEMPT.

Request Lenses

queryObjects_query :: Lens' QueryObjects (Maybe Query)

The query that defines the objects to be returned. The Query object can contain a maximum of ten selectors. The conditions in the query are limited to top-level String fields in the object. These filters can be applied to components, instances, and attempts.

queryObjects_limit :: Lens' QueryObjects (Maybe Int)

The maximum number of object names that QueryObjects will return in a single call. The default value is 100.

queryObjects_marker :: Lens' QueryObjects (Maybe Text)

The starting point for the results to be returned. For the first call, this value should be empty. As long as there are more results, continue to call QueryObjects with the marker value from the previous call to retrieve the next set of results.

queryObjects_pipelineId :: Lens' QueryObjects Text

The ID of the pipeline.

queryObjects_sphere :: Lens' QueryObjects Text

Indicates whether the query applies to components or instances. The possible values are: COMPONENT, INSTANCE, and ATTEMPT.

Destructuring the Response

data QueryObjectsResponse

Contains the output of QueryObjects.

See: newQueryObjectsResponse smart constructor.

Constructors

QueryObjectsResponse'

Fields

ids :: Maybe [Text]

The identifiers that match the query selectors.

hasMoreResults :: Maybe Bool

Indicates whether there are more results that can be obtained by a subsequent call.

marker :: Maybe Text

The starting point for the next page of results. To view the next page of results, call QueryObjects again with this marker value. If the value is null, there are no more results.

httpStatus :: Int

The response's http status code.

newQueryObjectsResponse

Arguments

:: Int
$sel:httpStatus:QueryObjectsResponse'

-> QueryObjectsResponse

Create a value of QueryObjectsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ids:QueryObjectsResponse', queryObjectsResponse_ids - The identifiers that match the query selectors.

$sel:hasMoreResults:QueryObjectsResponse', queryObjectsResponse_hasMoreResults - Indicates whether there are more results that can be obtained by a subsequent call.

$sel:marker:QueryObjects', queryObjectsResponse_marker - The starting point for the next page of results. To view the next page of results, call QueryObjects again with this marker value. If the value is null, there are no more results.

$sel:httpStatus:QueryObjectsResponse', queryObjectsResponse_httpStatus - The response's http status code.

Response Lenses

queryObjectsResponse_ids :: Lens' QueryObjectsResponse (Maybe [Text])

The identifiers that match the query selectors.

queryObjectsResponse_hasMoreResults :: Lens' QueryObjectsResponse (Maybe Bool)

Indicates whether there are more results that can be obtained by a subsequent call.

queryObjectsResponse_marker :: Lens' QueryObjectsResponse (Maybe Text)

The starting point for the next page of results. To view the next page of results, call QueryObjects again with this marker value. If the value is null, there are no more results.

queryObjectsResponse_httpStatus :: Lens' QueryObjectsResponse Int

The response's http status code.