Network.AWS.DataPipeline.Types

Description

Service Configuration

defaultService :: Service

API version 2012-10-29 of the Amazon Data Pipeline SDK configuration.

Errors

_TaskNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified task was not found.

_InvalidRequestException :: AsError a => Getting (First ServiceError) a ServiceError

The request was not valid. Verify that your request was properly formatted, that the signature was generated with the correct credentials, and that you haven't exceeded any of the service limits for your account.

_PipelineNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified pipeline was not found. Verify that you used the correct user and account identifiers.

_PipelineDeletedException :: AsError a => Getting (First ServiceError) a ServiceError

The specified pipeline has been deleted.

_InternalServiceError :: AsError a => Getting (First ServiceError) a ServiceError

An internal service error occurred.

OperatorType

newtype OperatorType

Constructors

OperatorType'

Fields

fromOperatorType :: Text

Bundled Patterns

pattern OperatorType_BETWEEN :: OperatorType

pattern OperatorType_EQ :: OperatorType

pattern OperatorType_GE :: OperatorType

pattern OperatorType_LE :: OperatorType

pattern OperatorType_REF_EQ :: OperatorType

TaskStatus

newtype TaskStatus

Constructors

TaskStatus'

Fields

fromTaskStatus :: Text

Bundled Patterns

pattern TaskStatus_FAILED :: TaskStatus

pattern TaskStatus_FALSE :: TaskStatus

pattern TaskStatus_FINISHED :: TaskStatus

Field

data Field

A key-value pair that describes a property of a pipeline object. The value is specified as either a string value (StringValue) or a reference to another object (RefValue) but not as both.

See: newField smart constructor.

Constructors

Field'

Fields

stringValue :: Maybe Text

The field value, expressed as a String.

refValue :: Maybe Text

The field value, expressed as the identifier of another object.

key :: Text

The field identifier.

newField

Arguments

:: Text
$sel:key:Field'

-> Field

Create a value of Field with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stringValue:Field', field_stringValue - The field value, expressed as a String.

$sel:refValue:Field', field_refValue - The field value, expressed as the identifier of another object.

$sel:key:Field', field_key - The field identifier.

field_stringValue :: Lens' Field (Maybe Text)

The field value, expressed as a String.

field_refValue :: Lens' Field (Maybe Text)

The field value, expressed as the identifier of another object.

field_key :: Lens' Field Text

The field identifier.

InstanceIdentity

data InstanceIdentity

Identity information for the EC2 instance that is hosting the task runner. You can get this value by calling a metadata URI from the EC2 instance. For more information, see Instance Metadata in the Amazon Elastic Compute Cloud User Guide. Passing in this value proves that your task runner is running on an EC2 instance, and ensures the proper AWS Data Pipeline service charges are applied to your pipeline.

See: newInstanceIdentity smart constructor.

Constructors

InstanceIdentity'

Fields

document :: Maybe Text

A description of an EC2 instance that is generated when the instance is launched and exposed to the instance via the instance metadata service in the form of a JSON representation of an object.

signature :: Maybe Text

A signature which can be used to verify the accuracy and authenticity of the information provided in the instance identity document.

newInstanceIdentity :: InstanceIdentity

Create a value of InstanceIdentity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:document:InstanceIdentity', instanceIdentity_document - A description of an EC2 instance that is generated when the instance is launched and exposed to the instance via the instance metadata service in the form of a JSON representation of an object.

$sel:signature:InstanceIdentity', instanceIdentity_signature - A signature which can be used to verify the accuracy and authenticity of the information provided in the instance identity document.

instanceIdentity_document :: Lens' InstanceIdentity (Maybe Text)

A description of an EC2 instance that is generated when the instance is launched and exposed to the instance via the instance metadata service in the form of a JSON representation of an object.

instanceIdentity_signature :: Lens' InstanceIdentity (Maybe Text)

A signature which can be used to verify the accuracy and authenticity of the information provided in the instance identity document.

Operator

data Operator

Contains a logical operation for comparing the value of a field with a specified value.

See: newOperator smart constructor.

Constructors

Operator'

Fields

values :: Maybe [Text]

The value that the actual field value will be compared with.

type' :: Maybe OperatorType

The logical operation to be performed: equal (EQ), equal reference (REF_EQ), less than or equal (LE), greater than or equal (GE), or between (BETWEEN). Equal reference (REF_EQ) can be used only with reference fields. The other comparison types can be used only with String fields. The comparison types you can use apply only to certain object fields, as detailed below.

The comparison operators EQ and REF_EQ act on the following fields:

name

@sphere

parent

@componentParent

@instanceParent

@status

@scheduledStartTime

@scheduledEndTime

@actualStartTime

@actualEndTime

The comparison operators GE, LE, and BETWEEN act on the following fields:

@scheduledStartTime

@scheduledEndTime

@actualStartTime

@actualEndTime

Note that fields beginning with the at sign (@) are read-only and set by the web service. When you name fields, you should choose names containing only alpha-numeric values, as symbols may be reserved by AWS Data Pipeline. User-defined fields that you add to a pipeline should prefix their name with the string "my".

newOperator :: Operator

Create a value of Operator with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:values:Operator', operator_values - The value that the actual field value will be compared with.

$sel:type':Operator', operator_type - The logical operation to be performed: equal (EQ), equal reference (REF_EQ), less than or equal (LE), greater than or equal (GE), or between (BETWEEN). Equal reference (REF_EQ) can be used only with reference fields. The other comparison types can be used only with String fields. The comparison types you can use apply only to certain object fields, as detailed below.

The comparison operators EQ and REF_EQ act on the following fields:

name

@sphere

parent

@componentParent

@instanceParent

@status

@scheduledStartTime

@scheduledEndTime

@actualStartTime

@actualEndTime

The comparison operators GE, LE, and BETWEEN act on the following fields:

@scheduledStartTime

@scheduledEndTime

@actualStartTime

@actualEndTime

Note that fields beginning with the at sign (@) are read-only and set by the web service. When you name fields, you should choose names containing only alpha-numeric values, as symbols may be reserved by AWS Data Pipeline. User-defined fields that you add to a pipeline should prefix their name with the string "my".

operator_values :: Lens' Operator (Maybe [Text])

The value that the actual field value will be compared with.

operator_type :: Lens' Operator (Maybe OperatorType)

The logical operation to be performed: equal (EQ), equal reference (REF_EQ), less than or equal (LE), greater than or equal (GE), or between (BETWEEN). Equal reference (REF_EQ) can be used only with reference fields. The other comparison types can be used only with String fields. The comparison types you can use apply only to certain object fields, as detailed below.

The comparison operators EQ and REF_EQ act on the following fields:

name

@sphere

parent

@componentParent

@instanceParent

@status

@scheduledStartTime

@scheduledEndTime

@actualStartTime

@actualEndTime

The comparison operators GE, LE, and BETWEEN act on the following fields:

@scheduledStartTime

@scheduledEndTime

@actualStartTime

@actualEndTime

Note that fields beginning with the at sign (@) are read-only and set by the web service. When you name fields, you should choose names containing only alpha-numeric values, as symbols may be reserved by AWS Data Pipeline. User-defined fields that you add to a pipeline should prefix their name with the string "my".

ParameterAttribute

data ParameterAttribute

The attributes allowed or specified with a parameter object.

See: newParameterAttribute smart constructor.

Constructors

ParameterAttribute'

Fields

key :: Text

The field identifier.

stringValue :: Text

The field value, expressed as a String.

newParameterAttribute

Arguments

:: Text
$sel:key:ParameterAttribute'

-> Text
$sel:stringValue:ParameterAttribute'

-> ParameterAttribute

Create a value of ParameterAttribute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:ParameterAttribute', parameterAttribute_key - The field identifier.

$sel:stringValue:ParameterAttribute', parameterAttribute_stringValue - The field value, expressed as a String.

parameterAttribute_key :: Lens' ParameterAttribute Text

The field identifier.

parameterAttribute_stringValue :: Lens' ParameterAttribute Text

The field value, expressed as a String.

ParameterObject

data ParameterObject

Contains information about a parameter object.

See: newParameterObject smart constructor.

Constructors

ParameterObject'

Fields

id :: Text

The ID of the parameter object.

attributes :: [ParameterAttribute]

The attributes of the parameter object.

newParameterObject

Arguments

:: Text
$sel:id:ParameterObject'

-> ParameterObject

Create a value of ParameterObject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:ParameterObject', parameterObject_id - The ID of the parameter object.

$sel:attributes:ParameterObject', parameterObject_attributes - The attributes of the parameter object.

parameterObject_id :: Lens' ParameterObject Text

The ID of the parameter object.

parameterObject_attributes :: Lens' ParameterObject [ParameterAttribute]

The attributes of the parameter object.

ParameterValue

data ParameterValue

A value or list of parameter values.

See: newParameterValue smart constructor.

Constructors

ParameterValue'

Fields

id :: Text

The ID of the parameter value.

stringValue :: Text

The field value, expressed as a String.

newParameterValue

Arguments

:: Text
$sel:id:ParameterValue'

-> Text
$sel:stringValue:ParameterValue'

-> ParameterValue

Create a value of ParameterValue with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:ParameterValue', parameterValue_id - The ID of the parameter value.

$sel:stringValue:ParameterValue', parameterValue_stringValue - The field value, expressed as a String.

parameterValue_id :: Lens' ParameterValue Text

The ID of the parameter value.

parameterValue_stringValue :: Lens' ParameterValue Text

The field value, expressed as a String.

PipelineDescription

data PipelineDescription

Contains pipeline metadata.

See: newPipelineDescription smart constructor.

Constructors

PipelineDescription'

Fields

tags :: Maybe [Tag]

A list of tags to associated with a pipeline. Tags let you control access to pipelines. For more information, see Controlling User Access to Pipelines in the AWS Data Pipeline Developer Guide.

description :: Maybe Text

Description of the pipeline.

pipelineId :: Text

The pipeline identifier that was assigned by AWS Data Pipeline. This is a string of the form df-297EG78HU43EEXAMPLE.

name :: Text

The name of the pipeline.

fields :: [Field]

A list of read-only fields that contain metadata about the pipeline: @userId, @accountId, and @pipelineState.

newPipelineDescription

Arguments

:: Text
$sel:pipelineId:PipelineDescription'

-> Text
$sel:name:PipelineDescription'

-> PipelineDescription

Create a value of PipelineDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:PipelineDescription', pipelineDescription_tags - A list of tags to associated with a pipeline. Tags let you control access to pipelines. For more information, see Controlling User Access to Pipelines in the AWS Data Pipeline Developer Guide.

$sel:description:PipelineDescription', pipelineDescription_description - Description of the pipeline.

$sel:pipelineId:PipelineDescription', pipelineDescription_pipelineId - The pipeline identifier that was assigned by AWS Data Pipeline. This is a string of the form df-297EG78HU43EEXAMPLE.

$sel:name:PipelineDescription', pipelineDescription_name - The name of the pipeline.

$sel:fields:PipelineDescription', pipelineDescription_fields - A list of read-only fields that contain metadata about the pipeline: @userId, @accountId, and @pipelineState.

pipelineDescription_tags :: Lens' PipelineDescription (Maybe [Tag])

A list of tags to associated with a pipeline. Tags let you control access to pipelines. For more information, see Controlling User Access to Pipelines in the AWS Data Pipeline Developer Guide.

pipelineDescription_description :: Lens' PipelineDescription (Maybe Text)

Description of the pipeline.

pipelineDescription_pipelineId :: Lens' PipelineDescription Text

The pipeline identifier that was assigned by AWS Data Pipeline. This is a string of the form df-297EG78HU43EEXAMPLE.

pipelineDescription_name :: Lens' PipelineDescription Text

The name of the pipeline.

pipelineDescription_fields :: Lens' PipelineDescription [Field]

A list of read-only fields that contain metadata about the pipeline: @userId, @accountId, and @pipelineState.

PipelineIdName

data PipelineIdName

Contains the name and identifier of a pipeline.

See: newPipelineIdName smart constructor.

Constructors

PipelineIdName'

Fields

id :: Maybe Text

The ID of the pipeline that was assigned by AWS Data Pipeline. This is a string of the form df-297EG78HU43EEXAMPLE.

name :: Maybe Text

The name of the pipeline.

newPipelineIdName :: PipelineIdName

Create a value of PipelineIdName with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:PipelineIdName', pipelineIdName_id - The ID of the pipeline that was assigned by AWS Data Pipeline. This is a string of the form df-297EG78HU43EEXAMPLE.

$sel:name:PipelineIdName', pipelineIdName_name - The name of the pipeline.

pipelineIdName_id :: Lens' PipelineIdName (Maybe Text)

The ID of the pipeline that was assigned by AWS Data Pipeline. This is a string of the form df-297EG78HU43EEXAMPLE.

pipelineIdName_name :: Lens' PipelineIdName (Maybe Text)

The name of the pipeline.

PipelineObject

data PipelineObject

Contains information about a pipeline object. This can be a logical, physical, or physical attempt pipeline object. The complete set of components of a pipeline defines the pipeline.

See: newPipelineObject smart constructor.

Constructors

PipelineObject'

Fields

id :: Text

The ID of the object.

name :: Text

The name of the object.

fields :: [Field]

Key-value pairs that define the properties of the object.

newPipelineObject

Arguments

:: Text
$sel:id:PipelineObject'

-> Text
$sel:name:PipelineObject'

-> PipelineObject

Create a value of PipelineObject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:PipelineObject', pipelineObject_id - The ID of the object.

$sel:name:PipelineObject', pipelineObject_name - The name of the object.

$sel:fields:PipelineObject', pipelineObject_fields - Key-value pairs that define the properties of the object.

pipelineObject_id :: Lens' PipelineObject Text

The ID of the object.

pipelineObject_name :: Lens' PipelineObject Text

The name of the object.

pipelineObject_fields :: Lens' PipelineObject [Field]

Key-value pairs that define the properties of the object.

Query

data Query

Defines the query to run against an object.

See: newQuery smart constructor.

Constructors

Query'

Fields

selectors :: Maybe [Selector]

List of selectors that define the query. An object must satisfy all of the selectors to match the query.

newQuery :: Query

Create a value of Query with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:selectors:Query', query_selectors - List of selectors that define the query. An object must satisfy all of the selectors to match the query.

query_selectors :: Lens' Query (Maybe [Selector])

List of selectors that define the query. An object must satisfy all of the selectors to match the query.

Selector

data Selector

A comparision that is used to determine whether a query should return this object.

See: newSelector smart constructor.

Constructors

Selector'

Fields

operator :: Maybe Operator

fieldName :: Maybe Text

The name of the field that the operator will be applied to. The field name is the "key" portion of the field definition in the pipeline definition syntax that is used by the AWS Data Pipeline API. If the field is not set on the object, the condition fails.

newSelector :: Selector

Create a value of Selector with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operator:Selector', selector_operator - Undocumented member.

$sel:fieldName:Selector', selector_fieldName - The name of the field that the operator will be applied to. The field name is the "key" portion of the field definition in the pipeline definition syntax that is used by the AWS Data Pipeline API. If the field is not set on the object, the condition fails.

selector_operator :: Lens' Selector (Maybe Operator)

Undocumented member.

selector_fieldName :: Lens' Selector (Maybe Text)

The name of the field that the operator will be applied to. The field name is the "key" portion of the field definition in the pipeline definition syntax that is used by the AWS Data Pipeline API. If the field is not set on the object, the condition fails.

Tag

data Tag

Tags are key/value pairs defined by a user and associated with a pipeline to control access. AWS Data Pipeline allows you to associate ten tags per pipeline. For more information, see Controlling User Access to Pipelines in the AWS Data Pipeline Developer Guide.

See: newTag smart constructor.

Constructors

Tag'

Fields

key :: Text

The key name of a tag defined by a user. For more information, see Controlling User Access to Pipelines in the AWS Data Pipeline Developer Guide.

value :: Text

The optional value portion of a tag defined by a user. For more information, see Controlling User Access to Pipelines in the AWS Data Pipeline Developer Guide.

newTag

Arguments

:: Text
$sel:key:Tag'

-> Text
$sel:value:Tag'

-> Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:Tag', tag_key - The key name of a tag defined by a user. For more information, see Controlling User Access to Pipelines in the AWS Data Pipeline Developer Guide.

$sel:value:Tag', tag_value - The optional value portion of a tag defined by a user. For more information, see Controlling User Access to Pipelines in the AWS Data Pipeline Developer Guide.

tag_key :: Lens' Tag Text

The key name of a tag defined by a user. For more information, see Controlling User Access to Pipelines in the AWS Data Pipeline Developer Guide.

tag_value :: Lens' Tag Text

The optional value portion of a tag defined by a user. For more information, see Controlling User Access to Pipelines in the AWS Data Pipeline Developer Guide.

TaskObject

data TaskObject

Contains information about a pipeline task that is assigned to a task runner.

See: newTaskObject smart constructor.

Constructors

TaskObject'

Fields

pipelineId :: Maybe Text

The ID of the pipeline that provided the task.

objects :: Maybe (HashMap Text PipelineObject)

Connection information for the location where the task runner will publish the output of the task.

taskId :: Maybe Text

An internal identifier for the task. This ID is passed to the SetTaskStatus and ReportTaskProgress actions.

attemptId :: Maybe Text

The ID of the pipeline task attempt object. AWS Data Pipeline uses this value to track how many times a task is attempted.

newTaskObject :: TaskObject

Create a value of TaskObject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pipelineId:TaskObject', taskObject_pipelineId - The ID of the pipeline that provided the task.

$sel:objects:TaskObject', taskObject_objects - Connection information for the location where the task runner will publish the output of the task.

$sel:taskId:TaskObject', taskObject_taskId - An internal identifier for the task. This ID is passed to the SetTaskStatus and ReportTaskProgress actions.

$sel:attemptId:TaskObject', taskObject_attemptId - The ID of the pipeline task attempt object. AWS Data Pipeline uses this value to track how many times a task is attempted.

taskObject_pipelineId :: Lens' TaskObject (Maybe Text)

The ID of the pipeline that provided the task.

taskObject_objects :: Lens' TaskObject (Maybe (HashMap Text PipelineObject))

Connection information for the location where the task runner will publish the output of the task.

taskObject_taskId :: Lens' TaskObject (Maybe Text)

An internal identifier for the task. This ID is passed to the SetTaskStatus and ReportTaskProgress actions.

taskObject_attemptId :: Lens' TaskObject (Maybe Text)

The ID of the pipeline task attempt object. AWS Data Pipeline uses this value to track how many times a task is attempted.

ValidationError

data ValidationError

Defines a validation error. Validation errors prevent pipeline activation. The set of validation errors that can be returned are defined by AWS Data Pipeline.

See: newValidationError smart constructor.

Constructors

ValidationError'

Fields

id :: Maybe Text

The identifier of the object that contains the validation error.

errors :: Maybe [Text]

A description of the validation error.

newValidationError :: ValidationError

Create a value of ValidationError with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:ValidationError', validationError_id - The identifier of the object that contains the validation error.

$sel:errors:ValidationError', validationError_errors - A description of the validation error.

validationError_id :: Lens' ValidationError (Maybe Text)

The identifier of the object that contains the validation error.

validationError_errors :: Lens' ValidationError (Maybe [Text])

A description of the validation error.

ValidationWarning

data ValidationWarning

Defines a validation warning. Validation warnings do not prevent pipeline activation. The set of validation warnings that can be returned are defined by AWS Data Pipeline.

See: newValidationWarning smart constructor.

Constructors

ValidationWarning'

Fields

warnings :: Maybe [Text]

A description of the validation warning.

id :: Maybe Text

The identifier of the object that contains the validation warning.

newValidationWarning :: ValidationWarning

Create a value of ValidationWarning with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:warnings:ValidationWarning', validationWarning_warnings - A description of the validation warning.

$sel:id:ValidationWarning', validationWarning_id - The identifier of the object that contains the validation warning.

validationWarning_warnings :: Lens' ValidationWarning (Maybe [Text])

A description of the validation warning.

validationWarning_id :: Lens' ValidationWarning (Maybe Text)

The identifier of the object that contains the validation warning.