Network.AWS.DataPipeline.V2012_10_29

Description

AWS Data Pipeline is a web service that you can use to automate the movement and transformation of data. With AWS Data Pipeline, you can define data-driven workflows, so that tasks can be dependent on the successful completion of previous tasks.

Documentation

data ActivatePipeline

Constructors

ActivatePipeline

Fields

_apiPipelineId :: Text

The identifier of the pipeline to activate.

apiPipelineId :: Iso' ActivatePipeline Text

data ActivatePipelineResponse

Constructors

ActivatePipelineResponse

createPipeline

Arguments

:: Text
_cpiUniqueId

-> Text
_cpiName

-> CreatePipeline

Minimum specification for a CreatePipeline request.

data CreatePipeline

Constructors

CreatePipeline

Fields

_cpiUniqueId :: Text

A unique identifier that you specify. This identifier is not the same as the pipeline identifier assigned by AWS Data Pipeline. You are responsible for defining the format and ensuring the uniqueness of this identifier. You use this parameter to ensure idempotency during repeated calls to CreatePipeline. For example, if the first call to CreatePipeline does not return a clear success, you can pass in the same unique identifier and pipeline name combination on a subsequent call to CreatePipeline. CreatePipeline ensures that if a pipeline already exists with the same name and unique identifier, a new pipeline will not be created. Instead, you'll receive the pipeline identifier from the previous attempt. The uniqueness of the name and unique identifier combination is scoped to the AWS account or IAM user credentials.

_cpiName :: Text

The name of the new pipeline. You can use the same name for multiple pipelines associated with your AWS account, because AWS Data Pipeline assigns each new pipeline a unique pipeline identifier.

_cpiDescription :: Maybe Text

The description of the new pipeline.

cpiUniqueId :: Lens' CreatePipeline Text

cpiName :: Lens' CreatePipeline Text

cpiDescription :: Lens' CreatePipeline (Maybe Text)

data CreatePipelineResponse

Constructors

CreatePipelineResponse

Fields

_cpoPipelineId :: Text

The ID that AWS Data Pipeline assigns the newly created pipeline. The ID is a string of the form: df-06372391ZG65EXAMPLE.

cpoPipelineId :: Iso' CreatePipelineResponse Text

data DeletePipeline

Constructors

DeletePipeline

Fields

_dpjPipelineId :: Text

The identifier of the pipeline to be deleted.

dpjPipelineId :: Iso' DeletePipeline Text

data DeletePipelineResponse

Constructors

DeletePipelineResponse

describeObjects

Arguments

:: Text
_doiPipelineId

-> [Text]
_doiObjectIds

-> DescribeObjects

Minimum specification for a DescribeObjects request.

data DescribeObjects

Constructors

DescribeObjects

Fields

_doiPipelineId :: Text

Identifier of the pipeline that contains the object definitions.

_doiObjectIds :: [Text]

Identifiers of the pipeline objects that contain the definitions to be described. You can pass as many as 25 identifiers in a single call to DescribeObjects.

_doiEvaluateExpressions :: Maybe Bool

Indicates whether any expressions in the object should be evaluated when the object descriptions are returned.

_doiMarker :: Maybe Text

The starting point for the results to be returned. The first time you call DescribeObjects, this value should be empty. As long as the action returns HasMoreResults as True, you can call DescribeObjects again and pass the marker value from the response to retrieve the next set of results.

doiPipelineId :: Lens' DescribeObjects Text

doiObjectIds :: Lens' DescribeObjects [Text]

doiMarker :: Lens' DescribeObjects (Maybe Text)

doiEvaluateExpressions :: Lens' DescribeObjects (Maybe Bool)

data DescribeObjectsResponse

Constructors

DescribeObjectsResponse

Fields

_dooPipelineObjects :: [PipelineObject]

An array of object definitions that are returned by the call to DescribeObjects.

_dooHasMoreResults :: Bool

If True, there are more pages of results to return.

_dooMarker :: Maybe Text

The starting point for the next page of results. To view the next page of results, call DescribeObjects again with this marker value.

dooPipelineObjects :: Lens' DescribeObjectsResponse [PipelineObject]

dooMarker :: Lens' DescribeObjectsResponse (Maybe Text)

dooHasMoreResults :: Lens' DescribeObjectsResponse Bool

data DescribePipelines

Constructors

DescribePipelines

Fields

_dpiPipelineIds :: [Text]

Identifiers of the pipelines to describe. You can pass as many as 25 identifiers in a single call to DescribePipelines. You can obtain pipeline identifiers by calling ListPipelines.

dpiPipelineIds :: Iso' DescribePipelines [Text]

data DescribePipelinesResponse

Constructors

DescribePipelinesResponse

Fields

_dpoPipelineDescriptionList :: [PipelineDescription]

An array of descriptions returned for the specified pipelines.

dpoPipelineDescriptionList :: Iso' DescribePipelinesResponse [PipelineDescription]

data EvaluateExpression

Constructors

EvaluateExpression

Fields

_eeiPipelineId :: Text

The identifier of the pipeline.

_eeiObjectId :: Text

The identifier of the object.

_eeiExpression :: Text

The expression to evaluate.

eeiPipelineId :: Lens' EvaluateExpression Text

eeiObjectId :: Lens' EvaluateExpression Text

eeiExpression :: Lens' EvaluateExpression Text

data EvaluateExpressionResponse

Constructors

EvaluateExpressionResponse

Fields

_eeoEvaluatedExpression :: Text

The evaluated expression.

eeoEvaluatedExpression :: Iso' EvaluateExpressionResponse Text

getPipelineDefinition

Arguments

:: Text
_gpdiPipelineId

-> GetPipelineDefinition

Minimum specification for a GetPipelineDefinition request.

data GetPipelineDefinition

Constructors

GetPipelineDefinition

Fields

_gpdiPipelineId :: Text

The identifier of the pipeline.

_gpdiVersion :: Maybe Text

The version of the pipeline definition to retrieve. This parameter accepts the values latest (default) and active. Where latest indicates the last definition saved to the pipeline and active indicates the last definition of the pipeline that was activated.

gpdiVersion :: Lens' GetPipelineDefinition (Maybe Text)

gpdiPipelineId :: Lens' GetPipelineDefinition Text

data GetPipelineDefinitionResponse

Constructors

GetPipelineDefinitionResponse

Fields

_gpdoPipelineObjects :: [PipelineObject]

An array of objects defined in the pipeline.

gpdoPipelineObjects :: Iso' GetPipelineDefinitionResponse [PipelineObject]

listPipelines :: ListPipelines

Minimum specification for a ListPipelines request.

data ListPipelines

Constructors

ListPipelines

Fields

_lpiMarker :: Maybe Text

The starting point for the results to be returned. The first time you call ListPipelines, this value should be empty. As long as the action returns HasMoreResults as True, you can call ListPipelines again and pass the marker value from the response to retrieve the next set of results.

lpiMarker :: Iso' ListPipelines (Maybe Text)

data ListPipelinesResponse

Constructors

ListPipelinesResponse

Fields

_lpoPipelineIdList :: [PipelineIdName]

A list of all the pipeline identifiers that your account has permission to access. If you require additional information about the pipelines, you can use these identifiers to call DescribePipelines and GetPipelineDefinition.

_lpoHasMoreResults :: Bool

If True, there are more results that can be obtained by a subsequent call to ListPipelines.

_lpoMarker :: Maybe Text

If not null, indicates the starting point for the set of pipeline identifiers that the next call to ListPipelines will retrieve. If null, there are no more pipeline identifiers.

lpoPipelineIdList :: Lens' ListPipelinesResponse [PipelineIdName]

lpoMarker :: Lens' ListPipelinesResponse (Maybe Text)

lpoHasMoreResults :: Lens' ListPipelinesResponse Bool

pollForTask

Arguments

:: Text
_pftiWorkerGroup

-> PollForTask

Minimum specification for a PollForTask request.

data PollForTask

Constructors

PollForTask

Fields

_pftiWorkerGroup :: Text

Indicates the type of task the task runner is configured to accept and process. The worker group is set as a field on objects in the pipeline when they are created. You can only specify a single value for workerGroup in the call to PollForTask. There are no wildcard values permitted in workerGroup, the string must be an exact, case-sensitive, match.

_pftiHostname :: Maybe Text

The public DNS name of the calling task runner.

_pftiInstanceIdentity :: Maybe InstanceIdentity

Identity information for the Amazon EC2 instance that is hosting the task runner. You can get this value by calling the URI, http:/169.254.169.254latestmeta-datainstance-id, from the EC2 instance. For more information, go to Instance Metadata in the Amazon Elastic Compute Cloud User Guide. Passing in this value proves that your task runner is running on an EC2 instance, and ensures the proper AWS Data Pipeline service charges are applied to your pipeline.

pftiWorkerGroup :: Lens' PollForTask Text

pftiInstanceIdentity :: Lens' PollForTask (Maybe InstanceIdentity)

pftiHostname :: Lens' PollForTask (Maybe Text)

data PollForTaskResponse

Constructors

PollForTaskResponse

Fields

_pftoTaskObject :: Maybe TaskObject

An instance of PollForTaskResult, which contains an instance of TaskObject. The returned object contains all the information needed to complete the task that is being assigned to the task runner. One of the fields returned in this object is taskId, which contains an identifier for the task being assigned. The calling task runner uses taskId in subsequent calls to ReportTaskProgress and SetTaskStatus.

pftoTaskObject :: Iso' PollForTaskResponse (Maybe TaskObject)

data PutPipelineDefinition

Constructors

PutPipelineDefinition

Fields

_ppdiPipelineId :: Text

The identifier of the pipeline to be configured.

_ppdiPipelineObjects :: [PipelineObject]

The objects that define the pipeline. These will overwrite the existing pipeline definition.

ppdiPipelineObjects :: Lens' PutPipelineDefinition [PipelineObject]

ppdiPipelineId :: Lens' PutPipelineDefinition Text

data PutPipelineDefinitionResponse

Constructors

PutPipelineDefinitionResponse

Fields

_ppdoErrored :: Bool

If True, there were validation errors. If errored is True, the pipeline definition is stored but cannot be activated until you correct the pipeline and call PutPipelineDefinition to commit the corrected pipeline.

_ppdoValidationErrors :: [ValidationError]

A list of the validation errors that are associated with the objects defined in pipelineObjects.

_ppdoValidationWarnings :: [ValidationWarning]

A list of the validation warnings that are associated with the objects defined in pipelineObjects.

ppdoValidationWarnings :: Lens' PutPipelineDefinitionResponse [ValidationWarning]

ppdoValidationErrors :: Lens' PutPipelineDefinitionResponse [ValidationError]

ppdoErrored :: Lens' PutPipelineDefinitionResponse Bool

queryObjects

Arguments

:: Text
_qoiPipelineId

-> Text
_qoiSphere

-> QueryObjects

Minimum specification for a QueryObjects request.

data QueryObjects

Constructors

QueryObjects

Fields

_qoiPipelineId :: Text

Identifier of the pipeline to be queried for object names.

_qoiSphere :: Text

Specifies whether the query applies to components or instances. Allowable values: COMPONENT, INSTANCE, ATTEMPT.

_qoiLimit :: Maybe Integer

Specifies the maximum number of object names that QueryObjects will return in a single call. The default value is 100.

_qoiQuery :: Maybe Query

Query that defines the objects to be returned. The Query object can contain a maximum of ten selectors. The conditions in the query are limited to top-level String fields in the object. These filters can be applied to components, instances, and attempts.

_qoiMarker :: Maybe Text

The starting point for the results to be returned. The first time you call QueryObjects, this value should be empty. As long as the action returns HasMoreResults as True, you can call QueryObjects again and pass the marker value from the response to retrieve the next set of results.

qoiSphere :: Lens' QueryObjects Text

qoiQuery :: Lens' QueryObjects (Maybe Query)

qoiPipelineId :: Lens' QueryObjects Text

qoiMarker :: Lens' QueryObjects (Maybe Text)

qoiLimit :: Lens' QueryObjects (Maybe Integer)

data QueryObjectsResponse

Constructors

QueryObjectsResponse

Fields

_qooHasMoreResults :: Bool

If True, there are more results that can be obtained by a subsequent call to QueryObjects.

_qooIds :: [Text]

A list of identifiers that match the query selectors.

_qooMarker :: Maybe Text

The starting point for the results to be returned. As long as the action returns HasMoreResults as True, you can call QueryObjects again and pass the marker value from the response to retrieve the next set of results.

qooMarker :: Lens' QueryObjectsResponse (Maybe Text)

qooIds :: Lens' QueryObjectsResponse [Text]

qooHasMoreResults :: Lens' QueryObjectsResponse Bool

data ReportTaskProgress

Constructors

ReportTaskProgress

Fields

_rtpiTaskId :: Text

Identifier of the task assigned to the task runner. This value is provided in the TaskObject that the service returns with the response for the PollForTask action.

rtpiTaskId :: Iso' ReportTaskProgress Text

data ReportTaskProgressResponse

Constructors

ReportTaskProgressResponse

Fields

_rtpoCanceled :: Bool

If True, the calling task runner should cancel processing of the task. The task runner does not need to call SetTaskStatus for canceled tasks.

rtpoCanceled :: Iso' ReportTaskProgressResponse Bool

reportTaskRunnerHeartbeat

Arguments

:: Text
_rtrhiTaskrunnerId

-> ReportTaskRunnerHeartbeat

Minimum specification for a ReportTaskRunnerHeartbeat request.

data ReportTaskRunnerHeartbeat

Constructors

ReportTaskRunnerHeartbeat

Fields

_rtrhiTaskrunnerId :: Text

The identifier of the task runner. This value should be unique across your AWS account. In the case of AWS Data Pipeline Task Runner launched on a resource managed by AWS Data Pipeline, the web service provides a unique identifier when it launches the application. If you have written a custom task runner, you should assign a unique identifier for the task runner.

_rtrhiHostname :: Maybe Text

The public DNS name of the calling task runner.

_rtrhiWorkerGroup :: Maybe Text

Indicates the type of task the task runner is configured to accept and process. The worker group is set as a field on objects in the pipeline when they are created. You can only specify a single value for workerGroup in the call to ReportTaskRunnerHeartbeat. There are no wildcard values permitted in workerGroup, the string must be an exact, case-sensitive, match.

rtrhiWorkerGroup :: Lens' ReportTaskRunnerHeartbeat (Maybe Text)

rtrhiTaskrunnerId :: Lens' ReportTaskRunnerHeartbeat Text

rtrhiHostname :: Lens' ReportTaskRunnerHeartbeat (Maybe Text)

data ReportTaskRunnerHeartbeatResponse

Constructors

ReportTaskRunnerHeartbeatResponse

Fields

_rtrhoTerminate :: Bool

Indicates whether the calling task runner should terminate. If True, the task runner that called ReportTaskRunnerHeartbeat should terminate.

rtrhoTerminate :: Iso' ReportTaskRunnerHeartbeatResponse Bool

data SetStatus

Constructors

SetStatus

Fields

_ssiPipelineId :: Text

Identifies the pipeline that contains the objects.

_ssiObjectIds :: [Text]

Identifies an array of objects. The corresponding objects can be either physical or components, but not a mix of both types.

_ssiStatus :: Text

Specifies the status to be set on all the objects in objectIds. For components, this can be either PAUSE or RESUME. For instances, this can be either CANCEL, RERUN, or MARK_FINISHED.

ssiStatus :: Lens' SetStatus Text

ssiPipelineId :: Lens' SetStatus Text

ssiObjectIds :: Lens' SetStatus [Text]

data SetStatusResponse

Constructors

SetStatusResponse

setTaskStatus

Arguments

:: Text
_stsiTaskId

-> TaskStatus
_stsiTaskStatus

-> SetTaskStatus

Minimum specification for a SetTaskStatus request.

data SetTaskStatus

Constructors

SetTaskStatus

Fields

_stsiTaskId :: Text

Identifies the task assigned to the task runner. This value is set in the TaskObject that is returned by the PollForTask action.

_stsiTaskStatus :: TaskStatus

If FINISHED, the task successfully completed. If FAILED the task ended unsuccessfully. The FALSE value is used by preconditions.

_stsiErrorMessage :: Maybe Text

If an error occurred during the task, this value specifies a text description of the error. This value is set on the physical attempt object. It is used to display error information to the user. The web service does not parse this value.

_stsiErrorStackTrace :: Maybe Text

If an error occurred during the task, this value specifies the stack trace associated with the error. This value is set on the physical attempt object. It is used to display error information to the user. The web service does not parse this value.

_stsiErrorId :: Maybe Text

If an error occurred during the task, this value specifies an id value that represents the error. This value is set on the physical attempt object. It is used to display error information to the user. It should not start with string Service_ which is reserved by the system.

stsiTaskStatus :: Lens' SetTaskStatus TaskStatus

stsiTaskId :: Lens' SetTaskStatus Text

stsiErrorStackTrace :: Lens' SetTaskStatus (Maybe Text)

stsiErrorMessage :: Lens' SetTaskStatus (Maybe Text)

stsiErrorId :: Lens' SetTaskStatus (Maybe Text)

data SetTaskStatusResponse

Constructors

SetTaskStatusResponse

data DataPipeline

Supported version (2012-10-29) of the AWS Data Pipeline service.

data OperatorType

The logical operation to be performed: equal (EQ), equal reference (REF_EQ), less than or equal (LE), greater than or equal (GE), or between (BETWEEN). Equal reference (REF_EQ) can be used only with reference fields. The other comparison types can be used only with String fields. The comparison types you can use apply only to certain object fields, as detailed below. The comparison operators EQ and REF_EQ act on the following fields: name sphere parent componentParent instanceParent status scheduledStartTime scheduledEndTime actualStartTime actualEndTime The comparison operators GE, LE, and BETWEEN act on the following fields: scheduledStartTime scheduledEndTime actualStartTime actualEndTime Note that fields beginning with the at sign (@) are read-only and set by the web service. When you name fields, you should choose names containing only alpha-numeric values, as symbols may be reserved by AWS Data Pipeline. User-defined fields that you add to a pipeline should prefix their name with the string "my".

Constructors

OperatorTypeBetween
BETWEEN

OperatorTypeEq
EQ

OperatorTypeGe
GE

OperatorTypeLe
LE

OperatorTypeRefEq
REF_EQ

data TaskStatus

If FINISHED, the task successfully completed. If FAILED the task ended unsuccessfully. The FALSE value is used by preconditions.

Constructors

TaskStatusFailed
FAILED

TaskStatusFalse
FALSE

TaskStatusFinished
FINISHED

newtype Query

Query that defines the objects to be returned. The Query object can contain a maximum of ten selectors. The conditions in the query are limited to top-level String fields in the object. These filters can be applied to components, instances, and attempts.

Constructors

Query

Fields

_qSelectors :: [Selector]

List of selectors that define the query. An object must satisfy all of the selectors to match the query.

data Field

A key-value pair that describes a property of a pipeline object. The value is specified as either a string value (StringValue) or a reference to another object (RefValue) but not as both.

Constructors

Field

Fields

_fRefValue :: Maybe Text

The field value, expressed as the identifier of another object.

_fStringValue :: Maybe Text

The field value, expressed as a String.

_fKey :: Text

The field identifier.

data InstanceIdentity

Identity information for the Amazon EC2 instance that is hosting the task runner. You can get this value by calling the URI, http:/169.254.169.254latestmeta-datainstance-id, from the EC2 instance. For more information, go to Instance Metadata in the Amazon Elastic Compute Cloud User Guide. Passing in this value proves that your task runner is running on an EC2 instance, and ensures the proper AWS Data Pipeline service charges are applied to your pipeline.

Constructors

InstanceIdentity

Fields

_ikSignature :: Maybe Text

A signature which can be used to verify the accuracy and authenticity of the information provided in the instance identity document.

_ikDocument :: Maybe Text

A description of an Amazon EC2 instance that is generated when the instance is launched and exposed to the instance via the instance metadata service in the form of a JSON representation of an object.

data Operator

Contains a logical operation for comparing the value of a field with a specified value.

Constructors

Operator

Fields

_oValues :: [Text]

The value that the actual field value will be compared with.

_oType :: Maybe OperatorType

The logical operation to be performed: equal (EQ), equal reference (REF_EQ), less than or equal (LE), greater than or equal (GE), or between (BETWEEN). Equal reference (REF_EQ) can be used only with reference fields. The other comparison types can be used only with String fields. The comparison types you can use apply only to certain object fields, as detailed below. The comparison operators EQ and REF_EQ act on the following fields: name sphere parent componentParent instanceParent status scheduledStartTime scheduledEndTime actualStartTime actualEndTime The comparison operators GE, LE, and BETWEEN act on the following fields: scheduledStartTime scheduledEndTime actualStartTime actualEndTime Note that fields beginning with the at sign (@) are read-only and set by the web service. When you name fields, you should choose names containing only alpha-numeric values, as symbols may be reserved by AWS Data Pipeline. User-defined fields that you add to a pipeline should prefix their name with the string "my".

data PipelineDescription

Contains pipeline metadata.

Constructors

PipelineDescription

Fields

_pdPipelineId :: Text

The pipeline identifier that was assigned by AWS Data Pipeline. This is a string of the form df-297EG78HU43EEXAMPLE.

_pdName :: Text

Name of the pipeline.

_pdDescription :: Maybe Text

Description of the pipeline.

_pdFields :: [Field]

A list of read-only fields that contain metadata about the pipeline: userId, accountId, and @pipelineState.

data PipelineIdName

Contains the name and identifier of a pipeline.

Constructors

PipelineIdName

Fields

_pinName :: Maybe Text

Name of the pipeline.

_pinId :: Maybe Text

Identifier of the pipeline that was assigned by AWS Data Pipeline. This is a string of the form df-297EG78HU43EEXAMPLE.

data PipelineObject

Contains information about a pipeline object. This can be a logical, physical, or physical attempt pipeline object. The complete set of components of a pipeline defines the pipeline.

Constructors

PipelineObject

Fields

_poName :: Text

Name of the object.

_poId :: Text

Identifier of the object.

_poFields :: [Field]

Key-value pairs that define the properties of the object.

data Selector

A comparision that is used to determine whether a query should return this object.

Constructors

Selector

Fields

_uOperator :: Maybe Operator

Contains a logical operation for comparing the value of a field with a specified value.

_uFieldName :: Maybe Text

The name of the field that the operator will be applied to. The field name is the "key" portion of the field definition in the pipeline definition syntax that is used by the AWS Data Pipeline API. If the field is not set on the object, the condition fails.

data TaskObject

An instance of PollForTaskResult, which contains an instance of TaskObject. The returned object contains all the information needed to complete the task that is being assigned to the task runner. One of the fields returned in this object is taskId, which contains an identifier for the task being assigned. The calling task runner uses taskId in subsequent calls to ReportTaskProgress and SetTaskStatus.

Constructors

TaskObject

Fields

_toPipelineId :: Maybe Text

Identifier of the pipeline that provided the task.

_toAttemptId :: Maybe Text

Identifier of the pipeline task attempt object. AWS Data Pipeline uses this value to track how many times a task is attempted.

_toTaskId :: Maybe Text

An internal identifier for the task. This ID is passed to the SetTaskStatus and ReportTaskProgress actions.

_toObjects :: HashMap Text PipelineObject

Connection information for the location where the task runner will publish the output of the task.

data ValidationError

Defines a validation error returned by PutPipelineDefinition or ValidatePipelineDefinition. Validation errors prevent pipeline activation. The set of validation errors that can be returned are defined by AWS Data Pipeline.

Constructors

ValidationError

Fields

_vfId :: Maybe Text

The identifier of the object that contains the validation error.

_vfErrors :: [Text]

A description of the validation error.

data ValidationWarning

Defines a validation warning returned by PutPipelineDefinition or ValidatePipelineDefinition. Validation warnings do not prevent pipeline activation. The set of validation warnings that can be returned are defined by AWS Data Pipeline.

Constructors

ValidationWarning

Fields

_vxWarnings :: [Text]

A description of the validation warning.

_vxId :: Maybe Text

The identifier of the object that contains the validation warning.

qSelectors :: Iso' Query [Selector]

query :: Iso' [Selector] Query

fStringValue :: Lens' Field (Maybe Text)

fRefValue :: Lens' Field (Maybe Text)

fKey :: Lens' Field Text

ikSignature :: Lens' InstanceIdentity (Maybe Text)

ikDocument :: Lens' InstanceIdentity (Maybe Text)

oValues :: Lens' Operator [Text]

oType :: Lens' Operator (Maybe OperatorType)

pdPipelineId :: Lens' PipelineDescription Text

pdName :: Lens' PipelineDescription Text

pdFields :: Lens' PipelineDescription [Field]

pdDescription :: Lens' PipelineDescription (Maybe Text)

pinName :: Lens' PipelineIdName (Maybe Text)

pinId :: Lens' PipelineIdName (Maybe Text)

poName :: Lens' PipelineObject Text

poId :: Lens' PipelineObject Text

poFields :: Lens' PipelineObject [Field]

uOperator :: Lens' Selector (Maybe Operator)

uFieldName :: Lens' Selector (Maybe Text)

toTaskId :: Lens' TaskObject (Maybe Text)

toPipelineId :: Lens' TaskObject (Maybe Text)

toObjects :: Lens' TaskObject (HashMap Text PipelineObject)

toAttemptId :: Lens' TaskObject (Maybe Text)

vfId :: Lens' ValidationError (Maybe Text)

vfErrors :: Lens' ValidationError [Text]

vxWarnings :: Lens' ValidationWarning [Text]

vxId :: Lens' ValidationWarning (Maybe Text)

data ValidatePipelineDefinition

Constructors

ValidatePipelineDefinition

Fields

_vpdiPipelineId :: Text

Identifies the pipeline whose definition is to be validated.

_vpdiPipelineObjects :: [PipelineObject]

A list of objects that define the pipeline changes to validate against the pipeline.

vpdiPipelineObjects :: Lens' ValidatePipelineDefinition [PipelineObject]

vpdiPipelineId :: Lens' ValidatePipelineDefinition Text

data ValidatePipelineDefinitionResponse

Constructors

ValidatePipelineDefinitionResponse

Fields

_vpdoErrored :: Bool

If True, there were validation errors.

_vpdoValidationErrors :: [ValidationError]

Lists the validation errors that were found by ValidatePipelineDefinition.

_vpdoValidationWarnings :: [ValidationWarning]

Lists the validation warnings that were found by ValidatePipelineDefinition.

vpdoValidationWarnings :: Lens' ValidatePipelineDefinitionResponse [ValidationWarning]

vpdoValidationErrors :: Lens' ValidatePipelineDefinitionResponse [ValidationError]

vpdoErrored :: Lens' ValidatePipelineDefinitionResponse Bool