Network.AWS.DataPipeline.ValidatePipelineDefinition

Description

Validates the specified pipeline definition to ensure that it is well formed and can be run without error.

Creating a Request

data ValidatePipelineDefinition

Contains the parameters for ValidatePipelineDefinition.

See: newValidatePipelineDefinition smart constructor.

Constructors

ValidatePipelineDefinition'

Fields

parameterValues :: Maybe [ParameterValue]

The parameter values used with the pipeline.

parameterObjects :: Maybe [ParameterObject]

The parameter objects used with the pipeline.

pipelineId :: Text

The ID of the pipeline.

pipelineObjects :: [PipelineObject]

The objects that define the pipeline changes to validate against the pipeline.

newValidatePipelineDefinition

Arguments

:: Text
$sel:pipelineId:ValidatePipelineDefinition'

-> ValidatePipelineDefinition

Create a value of ValidatePipelineDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parameterValues:ValidatePipelineDefinition', validatePipelineDefinition_parameterValues - The parameter values used with the pipeline.

$sel:parameterObjects:ValidatePipelineDefinition', validatePipelineDefinition_parameterObjects - The parameter objects used with the pipeline.

$sel:pipelineId:ValidatePipelineDefinition', validatePipelineDefinition_pipelineId - The ID of the pipeline.

$sel:pipelineObjects:ValidatePipelineDefinition', validatePipelineDefinition_pipelineObjects - The objects that define the pipeline changes to validate against the pipeline.

Request Lenses

validatePipelineDefinition_parameterValues :: Lens' ValidatePipelineDefinition (Maybe [ParameterValue])

The parameter values used with the pipeline.

validatePipelineDefinition_parameterObjects :: Lens' ValidatePipelineDefinition (Maybe [ParameterObject])

The parameter objects used with the pipeline.

validatePipelineDefinition_pipelineId :: Lens' ValidatePipelineDefinition Text

The ID of the pipeline.

validatePipelineDefinition_pipelineObjects :: Lens' ValidatePipelineDefinition [PipelineObject]

The objects that define the pipeline changes to validate against the pipeline.

Destructuring the Response

data ValidatePipelineDefinitionResponse

Contains the output of ValidatePipelineDefinition.

See: newValidatePipelineDefinitionResponse smart constructor.

Constructors

ValidatePipelineDefinitionResponse'

Fields

validationErrors :: Maybe [ValidationError]

Any validation errors that were found.

validationWarnings :: Maybe [ValidationWarning]

Any validation warnings that were found.

httpStatus :: Int

The response's http status code.

errored :: Bool

Indicates whether there were validation errors.

newValidatePipelineDefinitionResponse

Arguments

:: Int
$sel:httpStatus:ValidatePipelineDefinitionResponse'

-> Bool
$sel:errored:ValidatePipelineDefinitionResponse'

-> ValidatePipelineDefinitionResponse

Create a value of ValidatePipelineDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:validationErrors:ValidatePipelineDefinitionResponse', validatePipelineDefinitionResponse_validationErrors - Any validation errors that were found.

$sel:validationWarnings:ValidatePipelineDefinitionResponse', validatePipelineDefinitionResponse_validationWarnings - Any validation warnings that were found.

$sel:httpStatus:ValidatePipelineDefinitionResponse', validatePipelineDefinitionResponse_httpStatus - The response's http status code.

$sel:errored:ValidatePipelineDefinitionResponse', validatePipelineDefinitionResponse_errored - Indicates whether there were validation errors.

Response Lenses

validatePipelineDefinitionResponse_validationErrors :: Lens' ValidatePipelineDefinitionResponse (Maybe [ValidationError])

Any validation errors that were found.

validatePipelineDefinitionResponse_validationWarnings :: Lens' ValidatePipelineDefinitionResponse (Maybe [ValidationWarning])

Any validation warnings that were found.

validatePipelineDefinitionResponse_httpStatus :: Lens' ValidatePipelineDefinitionResponse Int

The response's http status code.

validatePipelineDefinitionResponse_errored :: Lens' ValidatePipelineDefinitionResponse Bool

Indicates whether there were validation errors.