Network.AWS.DataSync.CreateAgent

Description

Activates an DataSync agent that you have deployed on your host. The activation process associates your agent with your account. In the activation process, you specify information such as the Amazon Web Services Region that you want to activate the agent in. You activate the agent in the Amazon Web Services Region where your target locations (in Amazon S3 or Amazon EFS) reside. Your tasks are created in this Amazon Web Services Region.

You can activate the agent in a VPC (virtual private cloud) or provide the agent access to a VPC endpoint so you can run tasks without going over the public internet.

You can use an agent for more than one location. If a task uses multiple agents, all of them need to have status AVAILABLE for the task to run. If you use multiple agents for a source location, the status of all the agents must be AVAILABLE for the task to run.

Agents are automatically updated by Amazon Web Services on a regular basis, using a mechanism that ensures minimal interruption to your tasks.

Creating a Request

data CreateAgent

CreateAgentRequest

See: newCreateAgent smart constructor.

Constructors

CreateAgent'

Fields

securityGroupArns :: Maybe (NonEmpty Text)

The ARNs of the security groups used to protect your data transfer task subnets. See SecurityGroupArns.

subnetArns :: Maybe (NonEmpty Text)

The Amazon Resource Names (ARNs) of the subnets in which DataSync will create elastic network interfaces for each data transfer task. The agent that runs a task must be private. When you start a task that is associated with an agent created in a VPC, or one that has access to an IP address in a VPC, then the task is also private. In this case, DataSync creates four network interfaces for each task in your subnet. For a data transfer to work, the agent must be able to route to all these four network interfaces.

agentName :: Maybe Text

The name you configured for your agent. This value is a text reference that is used to identify the agent in the console.

vpcEndpointId :: Maybe Text

The ID of the VPC (virtual private cloud) endpoint that the agent has access to. This is the client-side VPC endpoint, also called a PrivateLink. If you don't have a PrivateLink VPC endpoint, see Creating a VPC Endpoint Service Configuration in the Amazon VPC User Guide.

VPC endpoint ID looks like this: vpce-01234d5aff67890e1.

tags :: Maybe [TagListEntry]

The key-value pair that represents the tag that you want to associate with the agent. The value can be an empty string. This value helps you manage, filter, and search for your agents.

Valid characters for key and value are letters, spaces, and numbers representable in UTF-8 format, and the following special characters: + - = . _ : / @.

activationKey :: Text

Your agent activation key. You can get the activation key either by sending an HTTP GET request with redirects that enable you to get the agent IP address (port 80). Alternatively, you can get it from the DataSync console.

The redirect URL returned in the response provides you the activation key for your agent in the query string parameter activationKey. It might also include other activation-related parameters; however, these are merely defaults. The arguments you pass to this API call determine the actual configuration of your agent.

For more information, see Activating an Agent in the /DataSync User Guide./

newCreateAgent

Arguments

:: Text
$sel:activationKey:CreateAgent'

-> CreateAgent

Create a value of CreateAgent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:securityGroupArns:CreateAgent', createAgent_securityGroupArns - The ARNs of the security groups used to protect your data transfer task subnets. See SecurityGroupArns.

$sel:subnetArns:CreateAgent', createAgent_subnetArns - The Amazon Resource Names (ARNs) of the subnets in which DataSync will create elastic network interfaces for each data transfer task. The agent that runs a task must be private. When you start a task that is associated with an agent created in a VPC, or one that has access to an IP address in a VPC, then the task is also private. In this case, DataSync creates four network interfaces for each task in your subnet. For a data transfer to work, the agent must be able to route to all these four network interfaces.

$sel:agentName:CreateAgent', createAgent_agentName - The name you configured for your agent. This value is a text reference that is used to identify the agent in the console.

$sel:vpcEndpointId:CreateAgent', createAgent_vpcEndpointId - The ID of the VPC (virtual private cloud) endpoint that the agent has access to. This is the client-side VPC endpoint, also called a PrivateLink. If you don't have a PrivateLink VPC endpoint, see Creating a VPC Endpoint Service Configuration in the Amazon VPC User Guide.

VPC endpoint ID looks like this: vpce-01234d5aff67890e1.

$sel:tags:CreateAgent', createAgent_tags - The key-value pair that represents the tag that you want to associate with the agent. The value can be an empty string. This value helps you manage, filter, and search for your agents.

Valid characters for key and value are letters, spaces, and numbers representable in UTF-8 format, and the following special characters: + - = . _ : / @.

$sel:activationKey:CreateAgent', createAgent_activationKey - Your agent activation key. You can get the activation key either by sending an HTTP GET request with redirects that enable you to get the agent IP address (port 80). Alternatively, you can get it from the DataSync console.

The redirect URL returned in the response provides you the activation key for your agent in the query string parameter activationKey. It might also include other activation-related parameters; however, these are merely defaults. The arguments you pass to this API call determine the actual configuration of your agent.

For more information, see Activating an Agent in the /DataSync User Guide./

Request Lenses

createAgent_securityGroupArns :: Lens' CreateAgent (Maybe (NonEmpty Text))

The ARNs of the security groups used to protect your data transfer task subnets. See SecurityGroupArns.

createAgent_subnetArns :: Lens' CreateAgent (Maybe (NonEmpty Text))

The Amazon Resource Names (ARNs) of the subnets in which DataSync will create elastic network interfaces for each data transfer task. The agent that runs a task must be private. When you start a task that is associated with an agent created in a VPC, or one that has access to an IP address in a VPC, then the task is also private. In this case, DataSync creates four network interfaces for each task in your subnet. For a data transfer to work, the agent must be able to route to all these four network interfaces.

createAgent_agentName :: Lens' CreateAgent (Maybe Text)

The name you configured for your agent. This value is a text reference that is used to identify the agent in the console.

createAgent_vpcEndpointId :: Lens' CreateAgent (Maybe Text)

The ID of the VPC (virtual private cloud) endpoint that the agent has access to. This is the client-side VPC endpoint, also called a PrivateLink. If you don't have a PrivateLink VPC endpoint, see Creating a VPC Endpoint Service Configuration in the Amazon VPC User Guide.

VPC endpoint ID looks like this: vpce-01234d5aff67890e1.

createAgent_tags :: Lens' CreateAgent (Maybe [TagListEntry])

The key-value pair that represents the tag that you want to associate with the agent. The value can be an empty string. This value helps you manage, filter, and search for your agents.

Valid characters for key and value are letters, spaces, and numbers representable in UTF-8 format, and the following special characters: + - = . _ : / @.

createAgent_activationKey :: Lens' CreateAgent Text

Your agent activation key. You can get the activation key either by sending an HTTP GET request with redirects that enable you to get the agent IP address (port 80). Alternatively, you can get it from the DataSync console.

The redirect URL returned in the response provides you the activation key for your agent in the query string parameter activationKey. It might also include other activation-related parameters; however, these are merely defaults. The arguments you pass to this API call determine the actual configuration of your agent.

For more information, see Activating an Agent in the /DataSync User Guide./

Destructuring the Response

data CreateAgentResponse

CreateAgentResponse

See: newCreateAgentResponse smart constructor.

Constructors

CreateAgentResponse'

Fields

agentArn :: Maybe Text

The Amazon Resource Name (ARN) of the agent. Use the ListAgents operation to return a list of agents for your account and Amazon Web Services Region.

httpStatus :: Int

The response's http status code.

newCreateAgentResponse

Arguments

:: Int
$sel:httpStatus:CreateAgentResponse'

-> CreateAgentResponse

Create a value of CreateAgentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:agentArn:CreateAgentResponse', createAgentResponse_agentArn - The Amazon Resource Name (ARN) of the agent. Use the ListAgents operation to return a list of agents for your account and Amazon Web Services Region.

$sel:httpStatus:CreateAgentResponse', createAgentResponse_httpStatus - The response's http status code.

Response Lenses

createAgentResponse_agentArn :: Lens' CreateAgentResponse (Maybe Text)

The Amazon Resource Name (ARN) of the agent. Use the ListAgents operation to return a list of agents for your account and Amazon Web Services Region.

createAgentResponse_httpStatus :: Lens' CreateAgentResponse Int

The response's http status code.