Network.AWS.DataSync.CreateLocationFsxWindows

Description

Creates an endpoint for an Amazon FSx for Windows File Server file system.

Creating a Request

data CreateLocationFsxWindows

See: newCreateLocationFsxWindows smart constructor.

Constructors

CreateLocationFsxWindows'

Fields

domain :: Maybe Text

The name of the Windows domain that the FSx for Windows File Server belongs to.

subdirectory :: Maybe Text

A subdirectory in the location’s path. This subdirectory in the Amazon FSx for Windows File Server file system is used to read data from the Amazon FSx for Windows File Server source location or write data to the FSx for Windows File Server destination.

tags :: Maybe [TagListEntry]

The key-value pair that represents a tag that you want to add to the resource. The value can be an empty string. This value helps you manage, filter, and search for your resources. We recommend that you create a name tag for your location.

fsxFilesystemArn :: Text

The Amazon Resource Name (ARN) for the FSx for Windows File Server file system.

securityGroupArns :: NonEmpty Text

The Amazon Resource Names (ARNs) of the security groups that are to use to configure the FSx for Windows File Server file system.

user :: Text

The user who has the permissions to access files and folders in the FSx for Windows File Server file system.

For information about choosing a user name that ensures sufficient permissions to files, folders, and metadata, see user.

password :: Sensitive Text

The password of the user who has the permissions to access files and folders in the FSx for Windows File Server file system.

newCreateLocationFsxWindows

Arguments

:: Text
$sel:fsxFilesystemArn:CreateLocationFsxWindows'

-> NonEmpty Text
$sel:securityGroupArns:CreateLocationFsxWindows'

-> Text
$sel:user:CreateLocationFsxWindows'

-> Text
$sel:password:CreateLocationFsxWindows'

-> CreateLocationFsxWindows

Create a value of CreateLocationFsxWindows with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domain:CreateLocationFsxWindows', createLocationFsxWindows_domain - The name of the Windows domain that the FSx for Windows File Server belongs to.

$sel:subdirectory:CreateLocationFsxWindows', createLocationFsxWindows_subdirectory - A subdirectory in the location’s path. This subdirectory in the Amazon FSx for Windows File Server file system is used to read data from the Amazon FSx for Windows File Server source location or write data to the FSx for Windows File Server destination.

$sel:tags:CreateLocationFsxWindows', createLocationFsxWindows_tags - The key-value pair that represents a tag that you want to add to the resource. The value can be an empty string. This value helps you manage, filter, and search for your resources. We recommend that you create a name tag for your location.

$sel:fsxFilesystemArn:CreateLocationFsxWindows', createLocationFsxWindows_fsxFilesystemArn - The Amazon Resource Name (ARN) for the FSx for Windows File Server file system.

$sel:securityGroupArns:CreateLocationFsxWindows', createLocationFsxWindows_securityGroupArns - The Amazon Resource Names (ARNs) of the security groups that are to use to configure the FSx for Windows File Server file system.

$sel:user:CreateLocationFsxWindows', createLocationFsxWindows_user - The user who has the permissions to access files and folders in the FSx for Windows File Server file system.

For information about choosing a user name that ensures sufficient permissions to files, folders, and metadata, see user.

$sel:password:CreateLocationFsxWindows', createLocationFsxWindows_password - The password of the user who has the permissions to access files and folders in the FSx for Windows File Server file system.

Request Lenses

createLocationFsxWindows_domain :: Lens' CreateLocationFsxWindows (Maybe Text)

The name of the Windows domain that the FSx for Windows File Server belongs to.

createLocationFsxWindows_subdirectory :: Lens' CreateLocationFsxWindows (Maybe Text)

A subdirectory in the location’s path. This subdirectory in the Amazon FSx for Windows File Server file system is used to read data from the Amazon FSx for Windows File Server source location or write data to the FSx for Windows File Server destination.

createLocationFsxWindows_tags :: Lens' CreateLocationFsxWindows (Maybe [TagListEntry])

The key-value pair that represents a tag that you want to add to the resource. The value can be an empty string. This value helps you manage, filter, and search for your resources. We recommend that you create a name tag for your location.

createLocationFsxWindows_fsxFilesystemArn :: Lens' CreateLocationFsxWindows Text

The Amazon Resource Name (ARN) for the FSx for Windows File Server file system.

createLocationFsxWindows_securityGroupArns :: Lens' CreateLocationFsxWindows (NonEmpty Text)

The Amazon Resource Names (ARNs) of the security groups that are to use to configure the FSx for Windows File Server file system.

createLocationFsxWindows_user :: Lens' CreateLocationFsxWindows Text

The user who has the permissions to access files and folders in the FSx for Windows File Server file system.

For information about choosing a user name that ensures sufficient permissions to files, folders, and metadata, see user.

createLocationFsxWindows_password :: Lens' CreateLocationFsxWindows Text

The password of the user who has the permissions to access files and folders in the FSx for Windows File Server file system.

Destructuring the Response

data CreateLocationFsxWindowsResponse

See: newCreateLocationFsxWindowsResponse smart constructor.

Constructors

CreateLocationFsxWindowsResponse'

Fields

locationArn :: Maybe Text

The Amazon Resource Name (ARN) of the FSx for Windows File Server file system location that is created.

httpStatus :: Int

The response's http status code.

newCreateLocationFsxWindowsResponse

Arguments

:: Int
$sel:httpStatus:CreateLocationFsxWindowsResponse'

-> CreateLocationFsxWindowsResponse

Create a value of CreateLocationFsxWindowsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:locationArn:CreateLocationFsxWindowsResponse', createLocationFsxWindowsResponse_locationArn - The Amazon Resource Name (ARN) of the FSx for Windows File Server file system location that is created.

$sel:httpStatus:CreateLocationFsxWindowsResponse', createLocationFsxWindowsResponse_httpStatus - The response's http status code.

Response Lenses

createLocationFsxWindowsResponse_locationArn :: Lens' CreateLocationFsxWindowsResponse (Maybe Text)

The Amazon Resource Name (ARN) of the FSx for Windows File Server file system location that is created.

createLocationFsxWindowsResponse_httpStatus :: Lens' CreateLocationFsxWindowsResponse Int

The response's http status code.