Network.AWS.DataSync.CreateLocationNfs

Description

Defines a file system on a Network File System (NFS) server that can be read from or written to.

Creating a Request

data CreateLocationNfs

CreateLocationNfsRequest

See: newCreateLocationNfs smart constructor.

Constructors

CreateLocationNfs'

Fields

mountOptions :: Maybe NfsMountOptions

The NFS mount options that DataSync can use to mount your NFS share.

tags :: Maybe [TagListEntry]

The key-value pair that represents the tag that you want to add to the location. The value can be an empty string. We recommend using tags to name your resources.

subdirectory :: Text

The subdirectory in the NFS file system that is used to read data from the NFS source location or write data to the NFS destination. The NFS path should be a path that's exported by the NFS server, or a subdirectory of that path. The path should be such that it can be mounted by other NFS clients in your network.

To see all the paths exported by your NFS server, run "showmount -e nfs-server-name" from an NFS client that has access to your server. You can specify any directory that appears in the results, and any subdirectory of that directory. Ensure that the NFS export is accessible without Kerberos authentication.

To transfer all the data in the folder you specified, DataSync needs to have permissions to read all the data. To ensure this, either configure the NFS export with no_root_squash, or ensure that the permissions for all of the files that you want DataSync allow read access for all users. Doing either enables the agent to read the files. For the agent to access directories, you must additionally enable all execute access.

If you are copying data to or from your Snowcone device, see NFS Server on Snowcone for more information.

For information about NFS export configuration, see 18.7. The /etc/exports Configuration File in the Red Hat Enterprise Linux documentation.

serverHostname :: Text

The name of the NFS server. This value is the IP address or Domain Name Service (DNS) name of the NFS server. An agent that is installed on-premises uses this host name to mount the NFS server in a network.

If you are copying data to or from your Snowcone device, see NFS Server on Snowcone for more information.

This name must either be DNS-compliant or must be an IP version 4 (IPv4) address.

onPremConfig :: OnPremConfig

Contains a list of Amazon Resource Names (ARNs) of agents that are used to connect to an NFS server.

If you are copying data to or from your Snowcone device, see NFS Server on Snowcone for more information.

newCreateLocationNfs

Arguments

:: Text
$sel:subdirectory:CreateLocationNfs'

-> Text
$sel:serverHostname:CreateLocationNfs'

-> OnPremConfig
$sel:onPremConfig:CreateLocationNfs'

-> CreateLocationNfs

Create a value of CreateLocationNfs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:mountOptions:CreateLocationNfs', createLocationNfs_mountOptions - The NFS mount options that DataSync can use to mount your NFS share.

$sel:tags:CreateLocationNfs', createLocationNfs_tags - The key-value pair that represents the tag that you want to add to the location. The value can be an empty string. We recommend using tags to name your resources.

$sel:subdirectory:CreateLocationNfs', createLocationNfs_subdirectory - The subdirectory in the NFS file system that is used to read data from the NFS source location or write data to the NFS destination. The NFS path should be a path that's exported by the NFS server, or a subdirectory of that path. The path should be such that it can be mounted by other NFS clients in your network.

To see all the paths exported by your NFS server, run "showmount -e nfs-server-name" from an NFS client that has access to your server. You can specify any directory that appears in the results, and any subdirectory of that directory. Ensure that the NFS export is accessible without Kerberos authentication.

To transfer all the data in the folder you specified, DataSync needs to have permissions to read all the data. To ensure this, either configure the NFS export with no_root_squash, or ensure that the permissions for all of the files that you want DataSync allow read access for all users. Doing either enables the agent to read the files. For the agent to access directories, you must additionally enable all execute access.

If you are copying data to or from your Snowcone device, see NFS Server on Snowcone for more information.

For information about NFS export configuration, see 18.7. The /etc/exports Configuration File in the Red Hat Enterprise Linux documentation.

$sel:serverHostname:CreateLocationNfs', createLocationNfs_serverHostname - The name of the NFS server. This value is the IP address or Domain Name Service (DNS) name of the NFS server. An agent that is installed on-premises uses this host name to mount the NFS server in a network.

If you are copying data to or from your Snowcone device, see NFS Server on Snowcone for more information.

This name must either be DNS-compliant or must be an IP version 4 (IPv4) address.

$sel:onPremConfig:CreateLocationNfs', createLocationNfs_onPremConfig - Contains a list of Amazon Resource Names (ARNs) of agents that are used to connect to an NFS server.

If you are copying data to or from your Snowcone device, see NFS Server on Snowcone for more information.

Request Lenses

createLocationNfs_mountOptions :: Lens' CreateLocationNfs (Maybe NfsMountOptions)

The NFS mount options that DataSync can use to mount your NFS share.

createLocationNfs_tags :: Lens' CreateLocationNfs (Maybe [TagListEntry])

The key-value pair that represents the tag that you want to add to the location. The value can be an empty string. We recommend using tags to name your resources.

createLocationNfs_subdirectory :: Lens' CreateLocationNfs Text

The subdirectory in the NFS file system that is used to read data from the NFS source location or write data to the NFS destination. The NFS path should be a path that's exported by the NFS server, or a subdirectory of that path. The path should be such that it can be mounted by other NFS clients in your network.

To see all the paths exported by your NFS server, run "showmount -e nfs-server-name" from an NFS client that has access to your server. You can specify any directory that appears in the results, and any subdirectory of that directory. Ensure that the NFS export is accessible without Kerberos authentication.

To transfer all the data in the folder you specified, DataSync needs to have permissions to read all the data. To ensure this, either configure the NFS export with no_root_squash, or ensure that the permissions for all of the files that you want DataSync allow read access for all users. Doing either enables the agent to read the files. For the agent to access directories, you must additionally enable all execute access.

If you are copying data to or from your Snowcone device, see NFS Server on Snowcone for more information.

For information about NFS export configuration, see 18.7. The /etc/exports Configuration File in the Red Hat Enterprise Linux documentation.

createLocationNfs_serverHostname :: Lens' CreateLocationNfs Text

The name of the NFS server. This value is the IP address or Domain Name Service (DNS) name of the NFS server. An agent that is installed on-premises uses this host name to mount the NFS server in a network.

If you are copying data to or from your Snowcone device, see NFS Server on Snowcone for more information.

This name must either be DNS-compliant or must be an IP version 4 (IPv4) address.

createLocationNfs_onPremConfig :: Lens' CreateLocationNfs OnPremConfig

Contains a list of Amazon Resource Names (ARNs) of agents that are used to connect to an NFS server.

If you are copying data to or from your Snowcone device, see NFS Server on Snowcone for more information.

Destructuring the Response

data CreateLocationNfsResponse

CreateLocationNfsResponse

See: newCreateLocationNfsResponse smart constructor.

Constructors

CreateLocationNfsResponse'

Fields

locationArn :: Maybe Text

The Amazon Resource Name (ARN) of the source NFS file system location that is created.

httpStatus :: Int

The response's http status code.

newCreateLocationNfsResponse

Arguments

:: Int
$sel:httpStatus:CreateLocationNfsResponse'

-> CreateLocationNfsResponse

Create a value of CreateLocationNfsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:locationArn:CreateLocationNfsResponse', createLocationNfsResponse_locationArn - The Amazon Resource Name (ARN) of the source NFS file system location that is created.

$sel:httpStatus:CreateLocationNfsResponse', createLocationNfsResponse_httpStatus - The response's http status code.

Response Lenses

createLocationNfsResponse_locationArn :: Lens' CreateLocationNfsResponse (Maybe Text)

The Amazon Resource Name (ARN) of the source NFS file system location that is created.

createLocationNfsResponse_httpStatus :: Lens' CreateLocationNfsResponse Int

The response's http status code.