Network.AWS.DataSync.CreateLocationS3

Description

Creates an endpoint for an Amazon S3 bucket.

For more information, see Create an Amazon S3 location in the DataSync User Guide.

Creating a Request

data CreateLocationS3

CreateLocationS3Request

See: newCreateLocationS3 smart constructor.

Constructors

CreateLocationS3'

Fields

agentArns :: Maybe (NonEmpty Text)

If you are using DataSync on an Amazon Web Services Outpost, specify the Amazon Resource Names (ARNs) of the DataSync agents deployed on your Outpost. For more information about launching a DataSync agent on an Amazon Web Services Outpost, see Deploy your DataSync agent on Outposts.

s3StorageClass :: Maybe S3StorageClass

The Amazon S3 storage class that you want to store your files in when this location is used as a task destination. For buckets in Amazon Web Services Regions, the storage class defaults to Standard. For buckets on Outposts, the storage class defaults to Amazon Web Services S3 Outposts.

For more information about S3 storage classes, see Amazon S3 Storage Classes. Some storage classes have behaviors that can affect your S3 storage cost. For detailed information, see Considerations when working with S3 storage classes in DataSync.

subdirectory :: Maybe Text

A subdirectory in the Amazon S3 bucket. This subdirectory in Amazon S3 is used to read data from the S3 source location or write data to the S3 destination.

tags :: Maybe [TagListEntry]

The key-value pair that represents the tag that you want to add to the location. The value can be an empty string. We recommend using tags to name your resources.

s3BucketArn :: Text

The ARN of the Amazon S3 bucket. If the bucket is on an Amazon Web Services Outpost, this must be an access point ARN.

s3Config :: S3Config

newCreateLocationS3

Arguments

:: Text
$sel:s3BucketArn:CreateLocationS3'

-> S3Config
$sel:s3Config:CreateLocationS3'

-> CreateLocationS3

Create a value of CreateLocationS3 with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:agentArns:CreateLocationS3', createLocationS3_agentArns - If you are using DataSync on an Amazon Web Services Outpost, specify the Amazon Resource Names (ARNs) of the DataSync agents deployed on your Outpost. For more information about launching a DataSync agent on an Amazon Web Services Outpost, see Deploy your DataSync agent on Outposts.

$sel:s3StorageClass:CreateLocationS3', createLocationS3_s3StorageClass - The Amazon S3 storage class that you want to store your files in when this location is used as a task destination. For buckets in Amazon Web Services Regions, the storage class defaults to Standard. For buckets on Outposts, the storage class defaults to Amazon Web Services S3 Outposts.

For more information about S3 storage classes, see Amazon S3 Storage Classes. Some storage classes have behaviors that can affect your S3 storage cost. For detailed information, see Considerations when working with S3 storage classes in DataSync.

$sel:subdirectory:CreateLocationS3', createLocationS3_subdirectory - A subdirectory in the Amazon S3 bucket. This subdirectory in Amazon S3 is used to read data from the S3 source location or write data to the S3 destination.

$sel:tags:CreateLocationS3', createLocationS3_tags - The key-value pair that represents the tag that you want to add to the location. The value can be an empty string. We recommend using tags to name your resources.

$sel:s3BucketArn:CreateLocationS3', createLocationS3_s3BucketArn - The ARN of the Amazon S3 bucket. If the bucket is on an Amazon Web Services Outpost, this must be an access point ARN.

$sel:s3Config:CreateLocationS3', createLocationS3_s3Config - Undocumented member.

Request Lenses

createLocationS3_agentArns :: Lens' CreateLocationS3 (Maybe (NonEmpty Text))

If you are using DataSync on an Amazon Web Services Outpost, specify the Amazon Resource Names (ARNs) of the DataSync agents deployed on your Outpost. For more information about launching a DataSync agent on an Amazon Web Services Outpost, see Deploy your DataSync agent on Outposts.

createLocationS3_s3StorageClass :: Lens' CreateLocationS3 (Maybe S3StorageClass)

The Amazon S3 storage class that you want to store your files in when this location is used as a task destination. For buckets in Amazon Web Services Regions, the storage class defaults to Standard. For buckets on Outposts, the storage class defaults to Amazon Web Services S3 Outposts.

For more information about S3 storage classes, see Amazon S3 Storage Classes. Some storage classes have behaviors that can affect your S3 storage cost. For detailed information, see Considerations when working with S3 storage classes in DataSync.

createLocationS3_subdirectory :: Lens' CreateLocationS3 (Maybe Text)

A subdirectory in the Amazon S3 bucket. This subdirectory in Amazon S3 is used to read data from the S3 source location or write data to the S3 destination.

createLocationS3_tags :: Lens' CreateLocationS3 (Maybe [TagListEntry])

The key-value pair that represents the tag that you want to add to the location. The value can be an empty string. We recommend using tags to name your resources.

createLocationS3_s3BucketArn :: Lens' CreateLocationS3 Text

The ARN of the Amazon S3 bucket. If the bucket is on an Amazon Web Services Outpost, this must be an access point ARN.

createLocationS3_s3Config :: Lens' CreateLocationS3 S3Config

Undocumented member.

Destructuring the Response

data CreateLocationS3Response

CreateLocationS3Response

See: newCreateLocationS3Response smart constructor.

Constructors

CreateLocationS3Response'

Fields

locationArn :: Maybe Text

The Amazon Resource Name (ARN) of the source Amazon S3 bucket location that is created.

httpStatus :: Int

The response's http status code.

newCreateLocationS3Response

Arguments

:: Int
$sel:httpStatus:CreateLocationS3Response'

-> CreateLocationS3Response

Create a value of CreateLocationS3Response with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:locationArn:CreateLocationS3Response', createLocationS3Response_locationArn - The Amazon Resource Name (ARN) of the source Amazon S3 bucket location that is created.

$sel:httpStatus:CreateLocationS3Response', createLocationS3Response_httpStatus - The response's http status code.

Response Lenses

createLocationS3Response_locationArn :: Lens' CreateLocationS3Response (Maybe Text)

The Amazon Resource Name (ARN) of the source Amazon S3 bucket location that is created.

createLocationS3Response_httpStatus :: Lens' CreateLocationS3Response Int

The response's http status code.