Network.AWS.DataSync.CreateLocationSmb

Description

Defines a file system on a Server Message Block (SMB) server that can be read from or written to.

Creating a Request

data CreateLocationSmb

CreateLocationSmbRequest

See: newCreateLocationSmb smart constructor.

Constructors

CreateLocationSmb'

Fields

domain :: Maybe Text

The name of the Windows domain that the SMB server belongs to.

mountOptions :: Maybe SmbMountOptions

The mount options used by DataSync to access the SMB server.

tags :: Maybe [TagListEntry]

The key-value pair that represents the tag that you want to add to the location. The value can be an empty string. We recommend using tags to name your resources.

subdirectory :: Text

The subdirectory in the SMB file system that is used to read data from the SMB source location or write data to the SMB destination. The SMB path should be a path that's exported by the SMB server, or a subdirectory of that path. The path should be such that it can be mounted by other SMB clients in your network.

Subdirectory must be specified with forward slashes. For example, /path/to/folder.

To transfer all the data in the folder you specified, DataSync needs to have permissions to mount the SMB share, as well as to access all the data in that share. To ensure this, either ensure that the user/password specified belongs to the user who can mount the share, and who has the appropriate permissions for all of the files and directories that you want DataSync to access, or use credentials of a member of the Backup Operators group to mount the share. Doing either enables the agent to access the data. For the agent to access directories, you must additionally enable all execute access.

serverHostname :: Text

The name of the SMB server. This value is the IP address or Domain Name Service (DNS) name of the SMB server. An agent that is installed on-premises uses this hostname to mount the SMB server in a network.

This name must either be DNS-compliant or must be an IP version 4 (IPv4) address.

user :: Text

The user who can mount the share, has the permissions to access files and folders in the SMB share.

For information about choosing a user name that ensures sufficient permissions to files, folders, and metadata, see user.

password :: Sensitive Text

The password of the user who can mount the share, has the permissions to access files and folders in the SMB share.

agentArns :: NonEmpty Text

The Amazon Resource Names (ARNs) of agents to use for a Simple Message Block (SMB) location.

newCreateLocationSmb

Arguments

:: Text
$sel:subdirectory:CreateLocationSmb'

-> Text
$sel:serverHostname:CreateLocationSmb'

-> Text
$sel:user:CreateLocationSmb'

-> Text
$sel:password:CreateLocationSmb'

-> NonEmpty Text
$sel:agentArns:CreateLocationSmb'

-> CreateLocationSmb

Create a value of CreateLocationSmb with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domain:CreateLocationSmb', createLocationSmb_domain - The name of the Windows domain that the SMB server belongs to.

$sel:mountOptions:CreateLocationSmb', createLocationSmb_mountOptions - The mount options used by DataSync to access the SMB server.

$sel:tags:CreateLocationSmb', createLocationSmb_tags - The key-value pair that represents the tag that you want to add to the location. The value can be an empty string. We recommend using tags to name your resources.

$sel:subdirectory:CreateLocationSmb', createLocationSmb_subdirectory - The subdirectory in the SMB file system that is used to read data from the SMB source location or write data to the SMB destination. The SMB path should be a path that's exported by the SMB server, or a subdirectory of that path. The path should be such that it can be mounted by other SMB clients in your network.

Subdirectory must be specified with forward slashes. For example, /path/to/folder.

To transfer all the data in the folder you specified, DataSync needs to have permissions to mount the SMB share, as well as to access all the data in that share. To ensure this, either ensure that the user/password specified belongs to the user who can mount the share, and who has the appropriate permissions for all of the files and directories that you want DataSync to access, or use credentials of a member of the Backup Operators group to mount the share. Doing either enables the agent to access the data. For the agent to access directories, you must additionally enable all execute access.

$sel:serverHostname:CreateLocationSmb', createLocationSmb_serverHostname - The name of the SMB server. This value is the IP address or Domain Name Service (DNS) name of the SMB server. An agent that is installed on-premises uses this hostname to mount the SMB server in a network.

This name must either be DNS-compliant or must be an IP version 4 (IPv4) address.

$sel:user:CreateLocationSmb', createLocationSmb_user - The user who can mount the share, has the permissions to access files and folders in the SMB share.

For information about choosing a user name that ensures sufficient permissions to files, folders, and metadata, see user.

$sel:password:CreateLocationSmb', createLocationSmb_password - The password of the user who can mount the share, has the permissions to access files and folders in the SMB share.

$sel:agentArns:CreateLocationSmb', createLocationSmb_agentArns - The Amazon Resource Names (ARNs) of agents to use for a Simple Message Block (SMB) location.

Request Lenses

createLocationSmb_domain :: Lens' CreateLocationSmb (Maybe Text)

The name of the Windows domain that the SMB server belongs to.

createLocationSmb_mountOptions :: Lens' CreateLocationSmb (Maybe SmbMountOptions)

The mount options used by DataSync to access the SMB server.

createLocationSmb_tags :: Lens' CreateLocationSmb (Maybe [TagListEntry])

The key-value pair that represents the tag that you want to add to the location. The value can be an empty string. We recommend using tags to name your resources.

createLocationSmb_subdirectory :: Lens' CreateLocationSmb Text

The subdirectory in the SMB file system that is used to read data from the SMB source location or write data to the SMB destination. The SMB path should be a path that's exported by the SMB server, or a subdirectory of that path. The path should be such that it can be mounted by other SMB clients in your network.

Subdirectory must be specified with forward slashes. For example, /path/to/folder.

To transfer all the data in the folder you specified, DataSync needs to have permissions to mount the SMB share, as well as to access all the data in that share. To ensure this, either ensure that the user/password specified belongs to the user who can mount the share, and who has the appropriate permissions for all of the files and directories that you want DataSync to access, or use credentials of a member of the Backup Operators group to mount the share. Doing either enables the agent to access the data. For the agent to access directories, you must additionally enable all execute access.

createLocationSmb_serverHostname :: Lens' CreateLocationSmb Text

The name of the SMB server. This value is the IP address or Domain Name Service (DNS) name of the SMB server. An agent that is installed on-premises uses this hostname to mount the SMB server in a network.

This name must either be DNS-compliant or must be an IP version 4 (IPv4) address.

createLocationSmb_user :: Lens' CreateLocationSmb Text

The user who can mount the share, has the permissions to access files and folders in the SMB share.

For information about choosing a user name that ensures sufficient permissions to files, folders, and metadata, see user.

createLocationSmb_password :: Lens' CreateLocationSmb Text

The password of the user who can mount the share, has the permissions to access files and folders in the SMB share.

createLocationSmb_agentArns :: Lens' CreateLocationSmb (NonEmpty Text)

The Amazon Resource Names (ARNs) of agents to use for a Simple Message Block (SMB) location.

Destructuring the Response

data CreateLocationSmbResponse

CreateLocationSmbResponse

See: newCreateLocationSmbResponse smart constructor.

Constructors

CreateLocationSmbResponse'

Fields

locationArn :: Maybe Text

The Amazon Resource Name (ARN) of the source SMB file system location that is created.

httpStatus :: Int

The response's http status code.

newCreateLocationSmbResponse

Arguments

:: Int
$sel:httpStatus:CreateLocationSmbResponse'

-> CreateLocationSmbResponse

Create a value of CreateLocationSmbResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:locationArn:CreateLocationSmbResponse', createLocationSmbResponse_locationArn - The Amazon Resource Name (ARN) of the source SMB file system location that is created.

$sel:httpStatus:CreateLocationSmbResponse', createLocationSmbResponse_httpStatus - The response's http status code.

Response Lenses

createLocationSmbResponse_locationArn :: Lens' CreateLocationSmbResponse (Maybe Text)

The Amazon Resource Name (ARN) of the source SMB file system location that is created.

createLocationSmbResponse_httpStatus :: Lens' CreateLocationSmbResponse Int

The response's http status code.