Network.AWS.DataSync.CreateTask

Description

Creates a task.

A task includes a source location and a destination location, and a configuration that specifies how data is transferred. A task always transfers data from the source location to the destination location. The configuration specifies options such as task scheduling, bandwidth limits, etc. A task is the complete definition of a data transfer.

When you create a task that transfers data between Amazon Web Services services in different Amazon Web Services Regions, one of the two locations that you specify must reside in the Region where DataSync is being used. The other location must be specified in a different Region.

You can transfer data between commercial Amazon Web Services Regions except for China, or between Amazon Web Services GovCloud (US) Regions.

When you use DataSync to copy files or objects between Amazon Web Services Regions, you pay for data transfer between Regions. This is billed as data transfer OUT from your source Region to your destination Region. For more information, see Data Transfer pricing.

Creating a Request

data CreateTask

CreateTaskRequest

See: newCreateTask smart constructor.

Constructors

CreateTask'

Fields

schedule :: Maybe TaskSchedule

Specifies a schedule used to periodically transfer files from a source to a destination location. The schedule should be specified in UTC time. For more information, see Scheduling your task.

includes :: Maybe [FilterRule]

A list of filter rules that determines which files to include when running a task. The pattern should contain a single filter string that consists of the patterns to include. The patterns are delimited by "|" (that is, a pipe). For example: "/folder1|/folder2"

name :: Maybe Text

The name of a task. This value is a text reference that is used to identify the task in the console.

excludes :: Maybe [FilterRule]

A list of filter rules that determines which files to exclude from a task. The list should contain a single filter string that consists of the patterns to exclude. The patterns are delimited by "|" (that is, a pipe), for example, "/folder1|/folder2".

options :: Maybe Options

The set of configuration options that control the behavior of a single execution of the task that occurs when you call StartTaskExecution. You can configure these options to preserve metadata such as user ID (UID) and group ID (GID), file permissions, data integrity verification, and so on.

For each individual task execution, you can override these options by specifying the OverrideOptions before starting the task execution. For more information, see the StartTaskExecution operation.

cloudWatchLogGroupArn :: Maybe Text

The Amazon Resource Name (ARN) of the Amazon CloudWatch log group that is used to monitor and log events in the task.

tags :: Maybe [TagListEntry]

The key-value pair that represents the tag that you want to add to the resource. The value can be an empty string.

sourceLocationArn :: Text

The Amazon Resource Name (ARN) of the source location for the task.

destinationLocationArn :: Text

The Amazon Resource Name (ARN) of an Amazon Web Services storage resource's location.

newCreateTask

Arguments

:: Text
$sel:sourceLocationArn:CreateTask'

-> Text
$sel:destinationLocationArn:CreateTask'

-> CreateTask

Create a value of CreateTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:schedule:CreateTask', createTask_schedule - Specifies a schedule used to periodically transfer files from a source to a destination location. The schedule should be specified in UTC time. For more information, see Scheduling your task.

$sel:includes:CreateTask', createTask_includes - A list of filter rules that determines which files to include when running a task. The pattern should contain a single filter string that consists of the patterns to include. The patterns are delimited by "|" (that is, a pipe). For example: "/folder1|/folder2"

$sel:name:CreateTask', createTask_name - The name of a task. This value is a text reference that is used to identify the task in the console.

$sel:excludes:CreateTask', createTask_excludes - A list of filter rules that determines which files to exclude from a task. The list should contain a single filter string that consists of the patterns to exclude. The patterns are delimited by "|" (that is, a pipe), for example, "/folder1|/folder2".

$sel:options:CreateTask', createTask_options - The set of configuration options that control the behavior of a single execution of the task that occurs when you call StartTaskExecution. You can configure these options to preserve metadata such as user ID (UID) and group ID (GID), file permissions, data integrity verification, and so on.

For each individual task execution, you can override these options by specifying the OverrideOptions before starting the task execution. For more information, see the StartTaskExecution operation.

$sel:cloudWatchLogGroupArn:CreateTask', createTask_cloudWatchLogGroupArn - The Amazon Resource Name (ARN) of the Amazon CloudWatch log group that is used to monitor and log events in the task.

$sel:tags:CreateTask', createTask_tags - The key-value pair that represents the tag that you want to add to the resource. The value can be an empty string.

$sel:sourceLocationArn:CreateTask', createTask_sourceLocationArn - The Amazon Resource Name (ARN) of the source location for the task.

$sel:destinationLocationArn:CreateTask', createTask_destinationLocationArn - The Amazon Resource Name (ARN) of an Amazon Web Services storage resource's location.

Request Lenses

createTask_schedule :: Lens' CreateTask (Maybe TaskSchedule)

Specifies a schedule used to periodically transfer files from a source to a destination location. The schedule should be specified in UTC time. For more information, see Scheduling your task.

createTask_includes :: Lens' CreateTask (Maybe [FilterRule])

A list of filter rules that determines which files to include when running a task. The pattern should contain a single filter string that consists of the patterns to include. The patterns are delimited by "|" (that is, a pipe). For example: "/folder1|/folder2"

createTask_name :: Lens' CreateTask (Maybe Text)

The name of a task. This value is a text reference that is used to identify the task in the console.

createTask_excludes :: Lens' CreateTask (Maybe [FilterRule])

A list of filter rules that determines which files to exclude from a task. The list should contain a single filter string that consists of the patterns to exclude. The patterns are delimited by "|" (that is, a pipe), for example, "/folder1|/folder2".

createTask_options :: Lens' CreateTask (Maybe Options)

The set of configuration options that control the behavior of a single execution of the task that occurs when you call StartTaskExecution. You can configure these options to preserve metadata such as user ID (UID) and group ID (GID), file permissions, data integrity verification, and so on.

For each individual task execution, you can override these options by specifying the OverrideOptions before starting the task execution. For more information, see the StartTaskExecution operation.

createTask_cloudWatchLogGroupArn :: Lens' CreateTask (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon CloudWatch log group that is used to monitor and log events in the task.

createTask_tags :: Lens' CreateTask (Maybe [TagListEntry])

The key-value pair that represents the tag that you want to add to the resource. The value can be an empty string.

createTask_sourceLocationArn :: Lens' CreateTask Text

The Amazon Resource Name (ARN) of the source location for the task.

createTask_destinationLocationArn :: Lens' CreateTask Text

The Amazon Resource Name (ARN) of an Amazon Web Services storage resource's location.

Destructuring the Response

data CreateTaskResponse

CreateTaskResponse

See: newCreateTaskResponse smart constructor.

Constructors

CreateTaskResponse'

Fields

taskArn :: Maybe Text

The Amazon Resource Name (ARN) of the task.

httpStatus :: Int

The response's http status code.

newCreateTaskResponse

Arguments

:: Int
$sel:httpStatus:CreateTaskResponse'

-> CreateTaskResponse

Create a value of CreateTaskResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:taskArn:CreateTaskResponse', createTaskResponse_taskArn - The Amazon Resource Name (ARN) of the task.

$sel:httpStatus:CreateTaskResponse', createTaskResponse_httpStatus - The response's http status code.

Response Lenses

createTaskResponse_taskArn :: Lens' CreateTaskResponse (Maybe Text)

The Amazon Resource Name (ARN) of the task.

createTaskResponse_httpStatus :: Lens' CreateTaskResponse Int

The response's http status code.