Network.AWS.DataSync.DescribeAgent

Description

Returns metadata such as the name, the network interfaces, and the status (that is, whether the agent is running or not) for an agent. To specify which agent to describe, use the Amazon Resource Name (ARN) of the agent in your request.

Creating a Request

data DescribeAgent

DescribeAgent

See: newDescribeAgent smart constructor.

Constructors

DescribeAgent'

Fields

agentArn :: Text

The Amazon Resource Name (ARN) of the agent to describe.

newDescribeAgent

Arguments

:: Text
$sel:agentArn:DescribeAgent'

-> DescribeAgent

Create a value of DescribeAgent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:agentArn:DescribeAgent', describeAgent_agentArn - The Amazon Resource Name (ARN) of the agent to describe.

Request Lenses

describeAgent_agentArn :: Lens' DescribeAgent Text

The Amazon Resource Name (ARN) of the agent to describe.

Destructuring the Response

data DescribeAgentResponse

DescribeAgentResponse

See: newDescribeAgentResponse smart constructor.

Constructors

DescribeAgentResponse'

Fields

creationTime :: Maybe POSIX

The time that the agent was activated (that is, created in your account).

status :: Maybe AgentStatus

The status of the agent. If the status is ONLINE, then the agent is configured properly and is available to use. The Running status is the normal running status for an agent. If the status is OFFLINE, the agent's VM is turned off or the agent is in an unhealthy state. When the issue that caused the unhealthy state is resolved, the agent returns to ONLINE status.

privateLinkConfig :: Maybe PrivateLinkConfig

The subnet and the security group that DataSync used to access a VPC endpoint.

endpointType :: Maybe EndpointType

The type of endpoint that your agent is connected to. If the endpoint is a VPC endpoint, the agent is not accessible over the public internet.

lastConnectionTime :: Maybe POSIX

The time that the agent last connected to DataSync.

agentArn :: Maybe Text

The Amazon Resource Name (ARN) of the agent.

name :: Maybe Text

The name of the agent.

httpStatus :: Int

The response's http status code.

newDescribeAgentResponse

Arguments

:: Int
$sel:httpStatus:DescribeAgentResponse'

-> DescribeAgentResponse

Create a value of DescribeAgentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:DescribeAgentResponse', describeAgentResponse_creationTime - The time that the agent was activated (that is, created in your account).

$sel:status:DescribeAgentResponse', describeAgentResponse_status - The status of the agent. If the status is ONLINE, then the agent is configured properly and is available to use. The Running status is the normal running status for an agent. If the status is OFFLINE, the agent's VM is turned off or the agent is in an unhealthy state. When the issue that caused the unhealthy state is resolved, the agent returns to ONLINE status.

$sel:privateLinkConfig:DescribeAgentResponse', describeAgentResponse_privateLinkConfig - The subnet and the security group that DataSync used to access a VPC endpoint.

$sel:endpointType:DescribeAgentResponse', describeAgentResponse_endpointType - The type of endpoint that your agent is connected to. If the endpoint is a VPC endpoint, the agent is not accessible over the public internet.

$sel:lastConnectionTime:DescribeAgentResponse', describeAgentResponse_lastConnectionTime - The time that the agent last connected to DataSync.

$sel:agentArn:DescribeAgent', describeAgentResponse_agentArn - The Amazon Resource Name (ARN) of the agent.

$sel:name:DescribeAgentResponse', describeAgentResponse_name - The name of the agent.

$sel:httpStatus:DescribeAgentResponse', describeAgentResponse_httpStatus - The response's http status code.

Response Lenses

describeAgentResponse_creationTime :: Lens' DescribeAgentResponse (Maybe UTCTime)

The time that the agent was activated (that is, created in your account).

describeAgentResponse_status :: Lens' DescribeAgentResponse (Maybe AgentStatus)

The status of the agent. If the status is ONLINE, then the agent is configured properly and is available to use. The Running status is the normal running status for an agent. If the status is OFFLINE, the agent's VM is turned off or the agent is in an unhealthy state. When the issue that caused the unhealthy state is resolved, the agent returns to ONLINE status.

describeAgentResponse_privateLinkConfig :: Lens' DescribeAgentResponse (Maybe PrivateLinkConfig)

The subnet and the security group that DataSync used to access a VPC endpoint.

describeAgentResponse_endpointType :: Lens' DescribeAgentResponse (Maybe EndpointType)

The type of endpoint that your agent is connected to. If the endpoint is a VPC endpoint, the agent is not accessible over the public internet.

describeAgentResponse_lastConnectionTime :: Lens' DescribeAgentResponse (Maybe UTCTime)

The time that the agent last connected to DataSync.

describeAgentResponse_agentArn :: Lens' DescribeAgentResponse (Maybe Text)

The Amazon Resource Name (ARN) of the agent.

describeAgentResponse_name :: Lens' DescribeAgentResponse (Maybe Text)

The name of the agent.

describeAgentResponse_httpStatus :: Lens' DescribeAgentResponse Int

The response's http status code.