Network.AWS.DataSync.DescribeLocationEfs

Description

Returns metadata, such as the path information about an Amazon EFS location.

Creating a Request

data DescribeLocationEfs

DescribeLocationEfsRequest

See: newDescribeLocationEfs smart constructor.

Constructors

DescribeLocationEfs'

Fields

locationArn :: Text

The Amazon Resource Name (ARN) of the EFS location to describe.

newDescribeLocationEfs

Arguments

:: Text
$sel:locationArn:DescribeLocationEfs'

-> DescribeLocationEfs

Create a value of DescribeLocationEfs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:locationArn:DescribeLocationEfs', describeLocationEfs_locationArn - The Amazon Resource Name (ARN) of the EFS location to describe.

Request Lenses

describeLocationEfs_locationArn :: Lens' DescribeLocationEfs Text

The Amazon Resource Name (ARN) of the EFS location to describe.

Destructuring the Response

data DescribeLocationEfsResponse

DescribeLocationEfsResponse

See: newDescribeLocationEfsResponse smart constructor.

Constructors

DescribeLocationEfsResponse'

Fields

creationTime :: Maybe POSIX

The time that the EFS location was created.

locationUri :: Maybe Text

The URL of the EFS location that was described.

locationArn :: Maybe Text

The Amazon Resource Name (ARN) of the EFS location that was described.

ec2Config :: Maybe Ec2Config

httpStatus :: Int

The response's http status code.

newDescribeLocationEfsResponse

Arguments

:: Int
$sel:httpStatus:DescribeLocationEfsResponse'

-> DescribeLocationEfsResponse

Create a value of DescribeLocationEfsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:DescribeLocationEfsResponse', describeLocationEfsResponse_creationTime - The time that the EFS location was created.

$sel:locationUri:DescribeLocationEfsResponse', describeLocationEfsResponse_locationUri - The URL of the EFS location that was described.

$sel:locationArn:DescribeLocationEfs', describeLocationEfsResponse_locationArn - The Amazon Resource Name (ARN) of the EFS location that was described.

$sel:ec2Config:DescribeLocationEfsResponse', describeLocationEfsResponse_ec2Config - Undocumented member.

$sel:httpStatus:DescribeLocationEfsResponse', describeLocationEfsResponse_httpStatus - The response's http status code.

Response Lenses

describeLocationEfsResponse_creationTime :: Lens' DescribeLocationEfsResponse (Maybe UTCTime)

The time that the EFS location was created.

describeLocationEfsResponse_locationUri :: Lens' DescribeLocationEfsResponse (Maybe Text)

The URL of the EFS location that was described.

describeLocationEfsResponse_locationArn :: Lens' DescribeLocationEfsResponse (Maybe Text)

The Amazon Resource Name (ARN) of the EFS location that was described.

describeLocationEfsResponse_ec2Config :: Lens' DescribeLocationEfsResponse (Maybe Ec2Config)

Undocumented member.

describeLocationEfsResponse_httpStatus :: Lens' DescribeLocationEfsResponse Int

The response's http status code.