Network.AWS.DataSync.DescribeLocationFsxWindows

Description

Returns metadata, such as the path information about an Amazon FSx for Windows File Server location.

Creating a Request

data DescribeLocationFsxWindows

See: newDescribeLocationFsxWindows smart constructor.

Constructors

DescribeLocationFsxWindows'

Fields

locationArn :: Text

The Amazon Resource Name (ARN) of the FSx for Windows File Server location to describe.

newDescribeLocationFsxWindows

Arguments

:: Text
$sel:locationArn:DescribeLocationFsxWindows'

-> DescribeLocationFsxWindows

Create a value of DescribeLocationFsxWindows with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:locationArn:DescribeLocationFsxWindows', describeLocationFsxWindows_locationArn - The Amazon Resource Name (ARN) of the FSx for Windows File Server location to describe.

Request Lenses

describeLocationFsxWindows_locationArn :: Lens' DescribeLocationFsxWindows Text

The Amazon Resource Name (ARN) of the FSx for Windows File Server location to describe.

Destructuring the Response

data DescribeLocationFsxWindowsResponse

See: newDescribeLocationFsxWindowsResponse smart constructor.

Constructors

DescribeLocationFsxWindowsResponse'

Fields

creationTime :: Maybe POSIX

The time that the FSx for Windows File Server location was created.

securityGroupArns :: Maybe (NonEmpty Text)

The Amazon Resource Names (ARNs) of the security groups that are configured for the FSx for Windows File Server file system.

domain :: Maybe Text

The name of the Windows domain that the FSx for Windows File Server belongs to.

locationUri :: Maybe Text

The URL of the FSx for Windows File Server location that was described.

user :: Maybe Text

The user who has the permissions to access files and folders in the FSx for Windows File Server file system.

locationArn :: Maybe Text

The Amazon Resource Name (ARN) of the FSx for Windows File Server location that was described.

httpStatus :: Int

The response's http status code.

newDescribeLocationFsxWindowsResponse

Arguments

:: Int
$sel:httpStatus:DescribeLocationFsxWindowsResponse'

-> DescribeLocationFsxWindowsResponse

Create a value of DescribeLocationFsxWindowsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:DescribeLocationFsxWindowsResponse', describeLocationFsxWindowsResponse_creationTime - The time that the FSx for Windows File Server location was created.

$sel:securityGroupArns:DescribeLocationFsxWindowsResponse', describeLocationFsxWindowsResponse_securityGroupArns - The Amazon Resource Names (ARNs) of the security groups that are configured for the FSx for Windows File Server file system.

$sel:domain:DescribeLocationFsxWindowsResponse', describeLocationFsxWindowsResponse_domain - The name of the Windows domain that the FSx for Windows File Server belongs to.

$sel:locationUri:DescribeLocationFsxWindowsResponse', describeLocationFsxWindowsResponse_locationUri - The URL of the FSx for Windows File Server location that was described.

$sel:user:DescribeLocationFsxWindowsResponse', describeLocationFsxWindowsResponse_user - The user who has the permissions to access files and folders in the FSx for Windows File Server file system.

$sel:locationArn:DescribeLocationFsxWindows', describeLocationFsxWindowsResponse_locationArn - The Amazon Resource Name (ARN) of the FSx for Windows File Server location that was described.

$sel:httpStatus:DescribeLocationFsxWindowsResponse', describeLocationFsxWindowsResponse_httpStatus - The response's http status code.

Response Lenses

describeLocationFsxWindowsResponse_creationTime :: Lens' DescribeLocationFsxWindowsResponse (Maybe UTCTime)

The time that the FSx for Windows File Server location was created.

describeLocationFsxWindowsResponse_securityGroupArns :: Lens' DescribeLocationFsxWindowsResponse (Maybe (NonEmpty Text))

The Amazon Resource Names (ARNs) of the security groups that are configured for the FSx for Windows File Server file system.

describeLocationFsxWindowsResponse_domain :: Lens' DescribeLocationFsxWindowsResponse (Maybe Text)

The name of the Windows domain that the FSx for Windows File Server belongs to.

describeLocationFsxWindowsResponse_locationUri :: Lens' DescribeLocationFsxWindowsResponse (Maybe Text)

The URL of the FSx for Windows File Server location that was described.

describeLocationFsxWindowsResponse_user :: Lens' DescribeLocationFsxWindowsResponse (Maybe Text)

The user who has the permissions to access files and folders in the FSx for Windows File Server file system.

describeLocationFsxWindowsResponse_locationArn :: Lens' DescribeLocationFsxWindowsResponse (Maybe Text)

The Amazon Resource Name (ARN) of the FSx for Windows File Server location that was described.

describeLocationFsxWindowsResponse_httpStatus :: Lens' DescribeLocationFsxWindowsResponse Int

The response's http status code.