Network.AWS.DataSync.DescribeLocationObjectStorage

Description

Returns metadata about a self-managed object storage server location. For more information about self-managed object storage locations, see Creating a location for object storage.

Creating a Request

data DescribeLocationObjectStorage

DescribeLocationObjectStorageRequest

See: newDescribeLocationObjectStorage smart constructor.

Constructors

DescribeLocationObjectStorage'

Fields

locationArn :: Text

The Amazon Resource Name (ARN) of the self-managed object storage server location that was described.

newDescribeLocationObjectStorage

Arguments

:: Text
$sel:locationArn:DescribeLocationObjectStorage'

-> DescribeLocationObjectStorage

Create a value of DescribeLocationObjectStorage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:locationArn:DescribeLocationObjectStorage', describeLocationObjectStorage_locationArn - The Amazon Resource Name (ARN) of the self-managed object storage server location that was described.

Request Lenses

describeLocationObjectStorage_locationArn :: Lens' DescribeLocationObjectStorage Text

The Amazon Resource Name (ARN) of the self-managed object storage server location that was described.

Destructuring the Response

data DescribeLocationObjectStorageResponse

DescribeLocationObjectStorageResponse

See: newDescribeLocationObjectStorageResponse smart constructor.

Constructors

DescribeLocationObjectStorageResponse'

Fields

serverPort :: Maybe Natural

The port that your self-managed object storage server accepts inbound network traffic on. The server port is set by default to TCP 80 (HTTP) or TCP 443 (HTTPS).

creationTime :: Maybe POSIX

The time that the self-managed object storage server agent was created.

agentArns :: Maybe (NonEmpty Text)

The Amazon Resource Name (ARN) of the agents associated with the self-managed object storage server location.

locationUri :: Maybe Text

The URL of the source self-managed object storage server location that was described.

serverProtocol :: Maybe ObjectStorageServerProtocol

The protocol that the object storage server uses to communicate. Valid values are HTTP or HTTPS.

locationArn :: Maybe Text

The Amazon Resource Name (ARN) of the self-managed object storage server location to describe.

accessKey :: Maybe Text

Optional. The access key is used if credentials are required to access the self-managed object storage server. If your object storage requires a user name and password to authenticate, use AccessKey and SecretKey to provide the user name and password, respectively.

httpStatus :: Int

The response's http status code.

newDescribeLocationObjectStorageResponse

Arguments

:: Int
$sel:httpStatus:DescribeLocationObjectStorageResponse'

-> DescribeLocationObjectStorageResponse

Create a value of DescribeLocationObjectStorageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serverPort:DescribeLocationObjectStorageResponse', describeLocationObjectStorageResponse_serverPort - The port that your self-managed object storage server accepts inbound network traffic on. The server port is set by default to TCP 80 (HTTP) or TCP 443 (HTTPS).

$sel:creationTime:DescribeLocationObjectStorageResponse', describeLocationObjectStorageResponse_creationTime - The time that the self-managed object storage server agent was created.

$sel:agentArns:DescribeLocationObjectStorageResponse', describeLocationObjectStorageResponse_agentArns - The Amazon Resource Name (ARN) of the agents associated with the self-managed object storage server location.

$sel:locationUri:DescribeLocationObjectStorageResponse', describeLocationObjectStorageResponse_locationUri - The URL of the source self-managed object storage server location that was described.

$sel:serverProtocol:DescribeLocationObjectStorageResponse', describeLocationObjectStorageResponse_serverProtocol - The protocol that the object storage server uses to communicate. Valid values are HTTP or HTTPS.

$sel:locationArn:DescribeLocationObjectStorage', describeLocationObjectStorageResponse_locationArn - The Amazon Resource Name (ARN) of the self-managed object storage server location to describe.

$sel:accessKey:DescribeLocationObjectStorageResponse', describeLocationObjectStorageResponse_accessKey - Optional. The access key is used if credentials are required to access the self-managed object storage server. If your object storage requires a user name and password to authenticate, use AccessKey and SecretKey to provide the user name and password, respectively.

$sel:httpStatus:DescribeLocationObjectStorageResponse', describeLocationObjectStorageResponse_httpStatus - The response's http status code.

Response Lenses

describeLocationObjectStorageResponse_serverPort :: Lens' DescribeLocationObjectStorageResponse (Maybe Natural)

The port that your self-managed object storage server accepts inbound network traffic on. The server port is set by default to TCP 80 (HTTP) or TCP 443 (HTTPS).

describeLocationObjectStorageResponse_creationTime :: Lens' DescribeLocationObjectStorageResponse (Maybe UTCTime)

The time that the self-managed object storage server agent was created.

describeLocationObjectStorageResponse_agentArns :: Lens' DescribeLocationObjectStorageResponse (Maybe (NonEmpty Text))

The Amazon Resource Name (ARN) of the agents associated with the self-managed object storage server location.

describeLocationObjectStorageResponse_locationUri :: Lens' DescribeLocationObjectStorageResponse (Maybe Text)

The URL of the source self-managed object storage server location that was described.

describeLocationObjectStorageResponse_serverProtocol :: Lens' DescribeLocationObjectStorageResponse (Maybe ObjectStorageServerProtocol)

The protocol that the object storage server uses to communicate. Valid values are HTTP or HTTPS.

describeLocationObjectStorageResponse_locationArn :: Lens' DescribeLocationObjectStorageResponse (Maybe Text)

The Amazon Resource Name (ARN) of the self-managed object storage server location to describe.

describeLocationObjectStorageResponse_accessKey :: Lens' DescribeLocationObjectStorageResponse (Maybe Text)

Optional. The access key is used if credentials are required to access the self-managed object storage server. If your object storage requires a user name and password to authenticate, use AccessKey and SecretKey to provide the user name and password, respectively.

describeLocationObjectStorageResponse_httpStatus :: Lens' DescribeLocationObjectStorageResponse Int

The response's http status code.