Network.AWS.DataSync.DescribeLocationSmb

Description

Returns metadata, such as the path and user information about an SMB location.

Creating a Request

data DescribeLocationSmb

DescribeLocationSmbRequest

See: newDescribeLocationSmb smart constructor.

Constructors

DescribeLocationSmb'

Fields

locationArn :: Text

The Amazon Resource Name (ARN) of the SMB location to describe.

newDescribeLocationSmb

Arguments

:: Text
$sel:locationArn:DescribeLocationSmb'

-> DescribeLocationSmb

Create a value of DescribeLocationSmb with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:locationArn:DescribeLocationSmb', describeLocationSmb_locationArn - The Amazon Resource Name (ARN) of the SMB location to describe.

Request Lenses

describeLocationSmb_locationArn :: Lens' DescribeLocationSmb Text

The Amazon Resource Name (ARN) of the SMB location to describe.

Destructuring the Response

data DescribeLocationSmbResponse

DescribeLocationSmbResponse

See: newDescribeLocationSmbResponse smart constructor.

Constructors

DescribeLocationSmbResponse'

Fields

creationTime :: Maybe POSIX

The time that the SMB location was created.

agentArns :: Maybe (NonEmpty Text)

The Amazon Resource Name (ARN) of the source SMB file system location that is created.

domain :: Maybe Text

The name of the Windows domain that the SMB server belongs to.

locationUri :: Maybe Text

The URL of the source SMB location that was described.

user :: Maybe Text

The user who can mount the share, has the permissions to access files and folders in the SMB share.

mountOptions :: Maybe SmbMountOptions

The mount options that are available for DataSync to use to access an SMB location.

locationArn :: Maybe Text

The Amazon Resource Name (ARN) of the SMB location that was described.

httpStatus :: Int

The response's http status code.

newDescribeLocationSmbResponse

Arguments

:: Int
$sel:httpStatus:DescribeLocationSmbResponse'

-> DescribeLocationSmbResponse

Create a value of DescribeLocationSmbResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:DescribeLocationSmbResponse', describeLocationSmbResponse_creationTime - The time that the SMB location was created.

$sel:agentArns:DescribeLocationSmbResponse', describeLocationSmbResponse_agentArns - The Amazon Resource Name (ARN) of the source SMB file system location that is created.

$sel:domain:DescribeLocationSmbResponse', describeLocationSmbResponse_domain - The name of the Windows domain that the SMB server belongs to.

$sel:locationUri:DescribeLocationSmbResponse', describeLocationSmbResponse_locationUri - The URL of the source SMB location that was described.

$sel:user:DescribeLocationSmbResponse', describeLocationSmbResponse_user - The user who can mount the share, has the permissions to access files and folders in the SMB share.

$sel:mountOptions:DescribeLocationSmbResponse', describeLocationSmbResponse_mountOptions - The mount options that are available for DataSync to use to access an SMB location.

$sel:locationArn:DescribeLocationSmb', describeLocationSmbResponse_locationArn - The Amazon Resource Name (ARN) of the SMB location that was described.

$sel:httpStatus:DescribeLocationSmbResponse', describeLocationSmbResponse_httpStatus - The response's http status code.

Response Lenses

describeLocationSmbResponse_creationTime :: Lens' DescribeLocationSmbResponse (Maybe UTCTime)

The time that the SMB location was created.

describeLocationSmbResponse_agentArns :: Lens' DescribeLocationSmbResponse (Maybe (NonEmpty Text))

The Amazon Resource Name (ARN) of the source SMB file system location that is created.

describeLocationSmbResponse_domain :: Lens' DescribeLocationSmbResponse (Maybe Text)

The name of the Windows domain that the SMB server belongs to.

describeLocationSmbResponse_locationUri :: Lens' DescribeLocationSmbResponse (Maybe Text)

The URL of the source SMB location that was described.

describeLocationSmbResponse_user :: Lens' DescribeLocationSmbResponse (Maybe Text)

The user who can mount the share, has the permissions to access files and folders in the SMB share.

describeLocationSmbResponse_mountOptions :: Lens' DescribeLocationSmbResponse (Maybe SmbMountOptions)

The mount options that are available for DataSync to use to access an SMB location.

describeLocationSmbResponse_locationArn :: Lens' DescribeLocationSmbResponse (Maybe Text)

The Amazon Resource Name (ARN) of the SMB location that was described.

describeLocationSmbResponse_httpStatus :: Lens' DescribeLocationSmbResponse Int

The response's http status code.