Network.AWS.DataSync.DescribeTask

Description

Returns metadata about a task.

Creating a Request

data DescribeTask

DescribeTaskRequest

See: newDescribeTask smart constructor.

Constructors

DescribeTask'

Fields

taskArn :: Text

The Amazon Resource Name (ARN) of the task to describe.

newDescribeTask

Arguments

:: Text
$sel:taskArn:DescribeTask'

-> DescribeTask

Create a value of DescribeTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:taskArn:DescribeTask', describeTask_taskArn - The Amazon Resource Name (ARN) of the task to describe.

Request Lenses

describeTask_taskArn :: Lens' DescribeTask Text

The Amazon Resource Name (ARN) of the task to describe.

Destructuring the Response

data DescribeTaskResponse

DescribeTaskResponse

See: newDescribeTaskResponse smart constructor.

Constructors

DescribeTaskResponse'

Fields

creationTime :: Maybe POSIX

The time that the task was created.

status :: Maybe TaskStatus

The status of the task that was described.

For detailed information about task execution statuses, see Understanding Task Statuses in the DataSync User Guide.

schedule :: Maybe TaskSchedule

The schedule used to periodically transfer files from a source to a destination location.

taskArn :: Maybe Text

The Amazon Resource Name (ARN) of the task that was described.

currentTaskExecutionArn :: Maybe Text

The Amazon Resource Name (ARN) of the task execution that is syncing files.

destinationNetworkInterfaceArns :: Maybe [Text]

The Amazon Resource Name (ARN) of the destination ENIs (Elastic Network Interface) that was created for your subnet.

destinationLocationArn :: Maybe Text

The Amazon Resource Name (ARN) of the Amazon Web Services storage resource's location.

includes :: Maybe [FilterRule]

A list of filter rules that determines which files to include when running a task. The pattern should contain a single filter string that consists of the patterns to include. The patterns are delimited by "|" (that is, a pipe). For example: "/folder1|/folder2"

name :: Maybe Text

The name of the task that was described.

errorCode :: Maybe Text

Errors that DataSync encountered during execution of the task. You can use this error code to help troubleshoot issues.

sourceLocationArn :: Maybe Text

The Amazon Resource Name (ARN) of the source file system's location.

excludes :: Maybe [FilterRule]

A list of filter rules that determines which files to exclude from a task. The list should contain a single filter string that consists of the patterns to exclude. The patterns are delimited by "|" (that is, a pipe), for example: "/folder1|/folder2"

sourceNetworkInterfaceArns :: Maybe [Text]

The Amazon Resource Name (ARN) of the source ENIs (Elastic Network Interface) that was created for your subnet.

options :: Maybe Options

The set of configuration options that control the behavior of a single execution of the task that occurs when you call StartTaskExecution. You can configure these options to preserve metadata such as user ID (UID) and group (GID), file permissions, data integrity verification, and so on.

For each individual task execution, you can override these options by specifying the overriding OverrideOptions value to StartTaskExecution operation.

cloudWatchLogGroupArn :: Maybe Text

The Amazon Resource Name (ARN) of the Amazon CloudWatch log group that was used to monitor and log events in the task.

For more information on these groups, see Working with Log Groups and Log Streams in the Amazon CloudWatch User Guide.

errorDetail :: Maybe Text

Detailed description of an error that was encountered during the task execution. You can use this information to help troubleshoot issues.

httpStatus :: Int

The response's http status code.

newDescribeTaskResponse

Arguments

:: Int
$sel:httpStatus:DescribeTaskResponse'

-> DescribeTaskResponse

Create a value of DescribeTaskResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:DescribeTaskResponse', describeTaskResponse_creationTime - The time that the task was created.

$sel:status:DescribeTaskResponse', describeTaskResponse_status - The status of the task that was described.

For detailed information about task execution statuses, see Understanding Task Statuses in the DataSync User Guide.

$sel:schedule:DescribeTaskResponse', describeTaskResponse_schedule - The schedule used to periodically transfer files from a source to a destination location.

$sel:taskArn:DescribeTask', describeTaskResponse_taskArn - The Amazon Resource Name (ARN) of the task that was described.

$sel:currentTaskExecutionArn:DescribeTaskResponse', describeTaskResponse_currentTaskExecutionArn - The Amazon Resource Name (ARN) of the task execution that is syncing files.

$sel:destinationNetworkInterfaceArns:DescribeTaskResponse', describeTaskResponse_destinationNetworkInterfaceArns - The Amazon Resource Name (ARN) of the destination ENIs (Elastic Network Interface) that was created for your subnet.

$sel:destinationLocationArn:DescribeTaskResponse', describeTaskResponse_destinationLocationArn - The Amazon Resource Name (ARN) of the Amazon Web Services storage resource's location.

$sel:includes:DescribeTaskResponse', describeTaskResponse_includes - A list of filter rules that determines which files to include when running a task. The pattern should contain a single filter string that consists of the patterns to include. The patterns are delimited by "|" (that is, a pipe). For example: "/folder1|/folder2"

$sel:name:DescribeTaskResponse', describeTaskResponse_name - The name of the task that was described.

$sel:errorCode:DescribeTaskResponse', describeTaskResponse_errorCode - Errors that DataSync encountered during execution of the task. You can use this error code to help troubleshoot issues.

$sel:sourceLocationArn:DescribeTaskResponse', describeTaskResponse_sourceLocationArn - The Amazon Resource Name (ARN) of the source file system's location.

$sel:excludes:DescribeTaskResponse', describeTaskResponse_excludes - A list of filter rules that determines which files to exclude from a task. The list should contain a single filter string that consists of the patterns to exclude. The patterns are delimited by "|" (that is, a pipe), for example: "/folder1|/folder2"

$sel:sourceNetworkInterfaceArns:DescribeTaskResponse', describeTaskResponse_sourceNetworkInterfaceArns - The Amazon Resource Name (ARN) of the source ENIs (Elastic Network Interface) that was created for your subnet.

$sel:options:DescribeTaskResponse', describeTaskResponse_options - The set of configuration options that control the behavior of a single execution of the task that occurs when you call StartTaskExecution. You can configure these options to preserve metadata such as user ID (UID) and group (GID), file permissions, data integrity verification, and so on.

For each individual task execution, you can override these options by specifying the overriding OverrideOptions value to StartTaskExecution operation.

$sel:cloudWatchLogGroupArn:DescribeTaskResponse', describeTaskResponse_cloudWatchLogGroupArn - The Amazon Resource Name (ARN) of the Amazon CloudWatch log group that was used to monitor and log events in the task.

For more information on these groups, see Working with Log Groups and Log Streams in the Amazon CloudWatch User Guide.

$sel:errorDetail:DescribeTaskResponse', describeTaskResponse_errorDetail - Detailed description of an error that was encountered during the task execution. You can use this information to help troubleshoot issues.

$sel:httpStatus:DescribeTaskResponse', describeTaskResponse_httpStatus - The response's http status code.

Response Lenses

describeTaskResponse_creationTime :: Lens' DescribeTaskResponse (Maybe UTCTime)

The time that the task was created.

describeTaskResponse_status :: Lens' DescribeTaskResponse (Maybe TaskStatus)

The status of the task that was described.

For detailed information about task execution statuses, see Understanding Task Statuses in the DataSync User Guide.

describeTaskResponse_schedule :: Lens' DescribeTaskResponse (Maybe TaskSchedule)

The schedule used to periodically transfer files from a source to a destination location.

describeTaskResponse_taskArn :: Lens' DescribeTaskResponse (Maybe Text)

The Amazon Resource Name (ARN) of the task that was described.

describeTaskResponse_currentTaskExecutionArn :: Lens' DescribeTaskResponse (Maybe Text)

The Amazon Resource Name (ARN) of the task execution that is syncing files.

describeTaskResponse_destinationNetworkInterfaceArns :: Lens' DescribeTaskResponse (Maybe [Text])

The Amazon Resource Name (ARN) of the destination ENIs (Elastic Network Interface) that was created for your subnet.

describeTaskResponse_destinationLocationArn :: Lens' DescribeTaskResponse (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon Web Services storage resource's location.

describeTaskResponse_includes :: Lens' DescribeTaskResponse (Maybe [FilterRule])

A list of filter rules that determines which files to include when running a task. The pattern should contain a single filter string that consists of the patterns to include. The patterns are delimited by "|" (that is, a pipe). For example: "/folder1|/folder2"

describeTaskResponse_name :: Lens' DescribeTaskResponse (Maybe Text)

The name of the task that was described.

describeTaskResponse_errorCode :: Lens' DescribeTaskResponse (Maybe Text)

Errors that DataSync encountered during execution of the task. You can use this error code to help troubleshoot issues.

describeTaskResponse_sourceLocationArn :: Lens' DescribeTaskResponse (Maybe Text)

The Amazon Resource Name (ARN) of the source file system's location.

describeTaskResponse_excludes :: Lens' DescribeTaskResponse (Maybe [FilterRule])

A list of filter rules that determines which files to exclude from a task. The list should contain a single filter string that consists of the patterns to exclude. The patterns are delimited by "|" (that is, a pipe), for example: "/folder1|/folder2"

describeTaskResponse_sourceNetworkInterfaceArns :: Lens' DescribeTaskResponse (Maybe [Text])

The Amazon Resource Name (ARN) of the source ENIs (Elastic Network Interface) that was created for your subnet.

describeTaskResponse_options :: Lens' DescribeTaskResponse (Maybe Options)

The set of configuration options that control the behavior of a single execution of the task that occurs when you call StartTaskExecution. You can configure these options to preserve metadata such as user ID (UID) and group (GID), file permissions, data integrity verification, and so on.

For each individual task execution, you can override these options by specifying the overriding OverrideOptions value to StartTaskExecution operation.

describeTaskResponse_cloudWatchLogGroupArn :: Lens' DescribeTaskResponse (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon CloudWatch log group that was used to monitor and log events in the task.

For more information on these groups, see Working with Log Groups and Log Streams in the Amazon CloudWatch User Guide.

describeTaskResponse_errorDetail :: Lens' DescribeTaskResponse (Maybe Text)

Detailed description of an error that was encountered during the task execution. You can use this information to help troubleshoot issues.

describeTaskResponse_httpStatus :: Lens' DescribeTaskResponse Int

The response's http status code.