Network.AWS.DataSync.DescribeTaskExecution

Description

Returns detailed metadata about a task that is being executed.

Creating a Request

data DescribeTaskExecution

DescribeTaskExecutionRequest

See: newDescribeTaskExecution smart constructor.

Constructors

DescribeTaskExecution'

Fields

taskExecutionArn :: Text

The Amazon Resource Name (ARN) of the task that is being executed.

newDescribeTaskExecution

Arguments

:: Text
$sel:taskExecutionArn:DescribeTaskExecution'

-> DescribeTaskExecution

Create a value of DescribeTaskExecution with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:taskExecutionArn:DescribeTaskExecution', describeTaskExecution_taskExecutionArn - The Amazon Resource Name (ARN) of the task that is being executed.

Request Lenses

describeTaskExecution_taskExecutionArn :: Lens' DescribeTaskExecution Text

The Amazon Resource Name (ARN) of the task that is being executed.

Destructuring the Response

data DescribeTaskExecutionResponse

DescribeTaskExecutionResponse

See: newDescribeTaskExecutionResponse smart constructor.

Constructors

DescribeTaskExecutionResponse'

Fields

status :: Maybe TaskExecutionStatus

The status of the task execution.

For detailed information about task execution statuses, see Understanding Task Statuses in the DataSync User Guide.

taskExecutionArn :: Maybe Text

The Amazon Resource Name (ARN) of the task execution that was described. TaskExecutionArn is hierarchical and includes TaskArn for the task that was executed.

For example, a TaskExecution value with the ARN arn:aws:datasync:us-east-1:111222333444:task/task-0208075f79cedf4a2/execution/exec-08ef1e88ec491019b executed the task with the ARN arn:aws:datasync:us-east-1:111222333444:task/task-0208075f79cedf4a2.

startTime :: Maybe POSIX

The time that the task execution was started.

filesTransferred :: Maybe Integer

The actual number of files that was transferred over the network. This value is calculated and updated on an ongoing basis during the TRANSFERRING phase. It's updated periodically when each file is read from the source and sent over the network.

If failures occur during a transfer, this value can be less than EstimatedFilesToTransfer. This value can also be greater than EstimatedFilesTransferred in some cases. This element is implementation-specific for some location types, so don't use it as an indicator for a correct file number or to monitor your task execution.

bytesWritten :: Maybe Integer

The number of logical bytes written to the destination Amazon Web Services storage resource.

bytesTransferred :: Maybe Integer

The physical number of bytes transferred over the network.

result :: Maybe TaskExecutionResultDetail

The result of the task execution.

includes :: Maybe [FilterRule]

A list of filter rules that determines which files to include when running a task. The list should contain a single filter string that consists of the patterns to include. The patterns are delimited by "|" (that is, a pipe), for example: "/folder1|/folder2"

estimatedFilesToTransfer :: Maybe Integer

The expected number of files that is to be transferred over the network. This value is calculated during the PREPARING phase, before the TRANSFERRING phase. This value is the expected number of files to be transferred. It's calculated based on comparing the content of the source and destination locations and finding the delta that needs to be transferred.

excludes :: Maybe [FilterRule]

A list of filter rules that determines which files to exclude from a task. The list should contain a single filter string that consists of the patterns to exclude. The patterns are delimited by "|" (that is, a pipe), for example: "/folder1|/folder2"

options :: Maybe Options

estimatedBytesToTransfer :: Maybe Integer

The estimated physical number of bytes that is to be transferred over the network.

httpStatus :: Int

The response's http status code.

newDescribeTaskExecutionResponse

Arguments

:: Int
$sel:httpStatus:DescribeTaskExecutionResponse'

-> DescribeTaskExecutionResponse

Create a value of DescribeTaskExecutionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DescribeTaskExecutionResponse', describeTaskExecutionResponse_status - The status of the task execution.

For detailed information about task execution statuses, see Understanding Task Statuses in the DataSync User Guide.

$sel:taskExecutionArn:DescribeTaskExecution', describeTaskExecutionResponse_taskExecutionArn - The Amazon Resource Name (ARN) of the task execution that was described. TaskExecutionArn is hierarchical and includes TaskArn for the task that was executed.

For example, a TaskExecution value with the ARN arn:aws:datasync:us-east-1:111222333444:task/task-0208075f79cedf4a2/execution/exec-08ef1e88ec491019b executed the task with the ARN arn:aws:datasync:us-east-1:111222333444:task/task-0208075f79cedf4a2.

$sel:startTime:DescribeTaskExecutionResponse', describeTaskExecutionResponse_startTime - The time that the task execution was started.

$sel:filesTransferred:DescribeTaskExecutionResponse', describeTaskExecutionResponse_filesTransferred - The actual number of files that was transferred over the network. This value is calculated and updated on an ongoing basis during the TRANSFERRING phase. It's updated periodically when each file is read from the source and sent over the network.

If failures occur during a transfer, this value can be less than EstimatedFilesToTransfer. This value can also be greater than EstimatedFilesTransferred in some cases. This element is implementation-specific for some location types, so don't use it as an indicator for a correct file number or to monitor your task execution.

$sel:bytesWritten:DescribeTaskExecutionResponse', describeTaskExecutionResponse_bytesWritten - The number of logical bytes written to the destination Amazon Web Services storage resource.

$sel:bytesTransferred:DescribeTaskExecutionResponse', describeTaskExecutionResponse_bytesTransferred - The physical number of bytes transferred over the network.

$sel:result:DescribeTaskExecutionResponse', describeTaskExecutionResponse_result - The result of the task execution.

$sel:includes:DescribeTaskExecutionResponse', describeTaskExecutionResponse_includes - A list of filter rules that determines which files to include when running a task. The list should contain a single filter string that consists of the patterns to include. The patterns are delimited by "|" (that is, a pipe), for example: "/folder1|/folder2"

$sel:estimatedFilesToTransfer:DescribeTaskExecutionResponse', describeTaskExecutionResponse_estimatedFilesToTransfer - The expected number of files that is to be transferred over the network. This value is calculated during the PREPARING phase, before the TRANSFERRING phase. This value is the expected number of files to be transferred. It's calculated based on comparing the content of the source and destination locations and finding the delta that needs to be transferred.

$sel:excludes:DescribeTaskExecutionResponse', describeTaskExecutionResponse_excludes - A list of filter rules that determines which files to exclude from a task. The list should contain a single filter string that consists of the patterns to exclude. The patterns are delimited by "|" (that is, a pipe), for example: "/folder1|/folder2"

$sel:options:DescribeTaskExecutionResponse', describeTaskExecutionResponse_options - Undocumented member.

$sel:estimatedBytesToTransfer:DescribeTaskExecutionResponse', describeTaskExecutionResponse_estimatedBytesToTransfer - The estimated physical number of bytes that is to be transferred over the network.

$sel:httpStatus:DescribeTaskExecutionResponse', describeTaskExecutionResponse_httpStatus - The response's http status code.

Response Lenses

describeTaskExecutionResponse_status :: Lens' DescribeTaskExecutionResponse (Maybe TaskExecutionStatus)

The status of the task execution.

For detailed information about task execution statuses, see Understanding Task Statuses in the DataSync User Guide.

describeTaskExecutionResponse_taskExecutionArn :: Lens' DescribeTaskExecutionResponse (Maybe Text)

The Amazon Resource Name (ARN) of the task execution that was described. TaskExecutionArn is hierarchical and includes TaskArn for the task that was executed.

For example, a TaskExecution value with the ARN arn:aws:datasync:us-east-1:111222333444:task/task-0208075f79cedf4a2/execution/exec-08ef1e88ec491019b executed the task with the ARN arn:aws:datasync:us-east-1:111222333444:task/task-0208075f79cedf4a2.

describeTaskExecutionResponse_startTime :: Lens' DescribeTaskExecutionResponse (Maybe UTCTime)

The time that the task execution was started.

describeTaskExecutionResponse_filesTransferred :: Lens' DescribeTaskExecutionResponse (Maybe Integer)

The actual number of files that was transferred over the network. This value is calculated and updated on an ongoing basis during the TRANSFERRING phase. It's updated periodically when each file is read from the source and sent over the network.

If failures occur during a transfer, this value can be less than EstimatedFilesToTransfer. This value can also be greater than EstimatedFilesTransferred in some cases. This element is implementation-specific for some location types, so don't use it as an indicator for a correct file number or to monitor your task execution.

describeTaskExecutionResponse_bytesWritten :: Lens' DescribeTaskExecutionResponse (Maybe Integer)

The number of logical bytes written to the destination Amazon Web Services storage resource.

describeTaskExecutionResponse_bytesTransferred :: Lens' DescribeTaskExecutionResponse (Maybe Integer)

The physical number of bytes transferred over the network.

describeTaskExecutionResponse_result :: Lens' DescribeTaskExecutionResponse (Maybe TaskExecutionResultDetail)

The result of the task execution.

describeTaskExecutionResponse_includes :: Lens' DescribeTaskExecutionResponse (Maybe [FilterRule])

A list of filter rules that determines which files to include when running a task. The list should contain a single filter string that consists of the patterns to include. The patterns are delimited by "|" (that is, a pipe), for example: "/folder1|/folder2"

describeTaskExecutionResponse_estimatedFilesToTransfer :: Lens' DescribeTaskExecutionResponse (Maybe Integer)

The expected number of files that is to be transferred over the network. This value is calculated during the PREPARING phase, before the TRANSFERRING phase. This value is the expected number of files to be transferred. It's calculated based on comparing the content of the source and destination locations and finding the delta that needs to be transferred.

describeTaskExecutionResponse_excludes :: Lens' DescribeTaskExecutionResponse (Maybe [FilterRule])

A list of filter rules that determines which files to exclude from a task. The list should contain a single filter string that consists of the patterns to exclude. The patterns are delimited by "|" (that is, a pipe), for example: "/folder1|/folder2"

describeTaskExecutionResponse_options :: Lens' DescribeTaskExecutionResponse (Maybe Options)

Undocumented member.

describeTaskExecutionResponse_estimatedBytesToTransfer :: Lens' DescribeTaskExecutionResponse (Maybe Integer)

The estimated physical number of bytes that is to be transferred over the network.

describeTaskExecutionResponse_httpStatus :: Lens' DescribeTaskExecutionResponse Int

The response's http status code.