Network.AWS.DataSync.Types.OnPremConfig

Description

Documentation

data OnPremConfig

A list of Amazon Resource Names (ARNs) of agents to use for a Network File System (NFS) location.

See: newOnPremConfig smart constructor.

Constructors

OnPremConfig'

Fields

agentArns :: NonEmpty Text

ARNs of the agents to use for an NFS location.

newOnPremConfig

Arguments

:: NonEmpty Text
$sel:agentArns:OnPremConfig'

-> OnPremConfig

Create a value of OnPremConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:agentArns:OnPremConfig', onPremConfig_agentArns - ARNs of the agents to use for an NFS location.

onPremConfig_agentArns :: Lens' OnPremConfig (NonEmpty Text)

ARNs of the agents to use for an NFS location.