Network.AWS.DataSync.Types.TaskExecutionResultDetail

Description

Documentation

data TaskExecutionResultDetail

Describes the detailed result of a TaskExecution operation. This result includes the time in milliseconds spent in each phase, the status of the task execution, and the errors encountered.

See: newTaskExecutionResultDetail smart constructor.

Constructors

TaskExecutionResultDetail'

Fields

prepareDuration :: Maybe Natural

The total time in milliseconds that DataSync spent in the PREPARING phase.

prepareStatus :: Maybe PhaseStatus

The status of the PREPARING phase.

verifyStatus :: Maybe PhaseStatus

The status of the VERIFYING phase.

verifyDuration :: Maybe Natural

The total time in milliseconds that DataSync spent in the VERIFYING phase.

totalDuration :: Maybe Natural

The total time in milliseconds that DataSync took to transfer the file from the source to the destination location.

transferStatus :: Maybe PhaseStatus

The status of the TRANSFERRING phase.

errorCode :: Maybe Text

Errors that DataSync encountered during execution of the task. You can use this error code to help troubleshoot issues.

transferDuration :: Maybe Natural

The total time in milliseconds that DataSync spent in the TRANSFERRING phase.

errorDetail :: Maybe Text

Detailed description of an error that was encountered during the task execution. You can use this information to help troubleshoot issues.

newTaskExecutionResultDetail :: TaskExecutionResultDetail

Create a value of TaskExecutionResultDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:prepareDuration:TaskExecutionResultDetail', taskExecutionResultDetail_prepareDuration - The total time in milliseconds that DataSync spent in the PREPARING phase.

$sel:prepareStatus:TaskExecutionResultDetail', taskExecutionResultDetail_prepareStatus - The status of the PREPARING phase.

$sel:verifyStatus:TaskExecutionResultDetail', taskExecutionResultDetail_verifyStatus - The status of the VERIFYING phase.

$sel:verifyDuration:TaskExecutionResultDetail', taskExecutionResultDetail_verifyDuration - The total time in milliseconds that DataSync spent in the VERIFYING phase.

$sel:totalDuration:TaskExecutionResultDetail', taskExecutionResultDetail_totalDuration - The total time in milliseconds that DataSync took to transfer the file from the source to the destination location.

$sel:transferStatus:TaskExecutionResultDetail', taskExecutionResultDetail_transferStatus - The status of the TRANSFERRING phase.

$sel:errorCode:TaskExecutionResultDetail', taskExecutionResultDetail_errorCode - Errors that DataSync encountered during execution of the task. You can use this error code to help troubleshoot issues.

$sel:transferDuration:TaskExecutionResultDetail', taskExecutionResultDetail_transferDuration - The total time in milliseconds that DataSync spent in the TRANSFERRING phase.

$sel:errorDetail:TaskExecutionResultDetail', taskExecutionResultDetail_errorDetail - Detailed description of an error that was encountered during the task execution. You can use this information to help troubleshoot issues.

taskExecutionResultDetail_prepareDuration :: Lens' TaskExecutionResultDetail (Maybe Natural)

The total time in milliseconds that DataSync spent in the PREPARING phase.

taskExecutionResultDetail_prepareStatus :: Lens' TaskExecutionResultDetail (Maybe PhaseStatus)

The status of the PREPARING phase.

taskExecutionResultDetail_verifyStatus :: Lens' TaskExecutionResultDetail (Maybe PhaseStatus)

The status of the VERIFYING phase.

taskExecutionResultDetail_verifyDuration :: Lens' TaskExecutionResultDetail (Maybe Natural)

The total time in milliseconds that DataSync spent in the VERIFYING phase.

taskExecutionResultDetail_totalDuration :: Lens' TaskExecutionResultDetail (Maybe Natural)

The total time in milliseconds that DataSync took to transfer the file from the source to the destination location.

taskExecutionResultDetail_transferStatus :: Lens' TaskExecutionResultDetail (Maybe PhaseStatus)

The status of the TRANSFERRING phase.

taskExecutionResultDetail_errorCode :: Lens' TaskExecutionResultDetail (Maybe Text)

Errors that DataSync encountered during execution of the task. You can use this error code to help troubleshoot issues.

taskExecutionResultDetail_transferDuration :: Lens' TaskExecutionResultDetail (Maybe Natural)

The total time in milliseconds that DataSync spent in the TRANSFERRING phase.

taskExecutionResultDetail_errorDetail :: Lens' TaskExecutionResultDetail (Maybe Text)

Detailed description of an error that was encountered during the task execution. You can use this information to help troubleshoot issues.