Network.AWS.DataSync.UpdateLocationObjectStorage

Description

Updates some of the parameters of a previously created location for self-managed object storage server access. For information about creating a self-managed object storage location, see Creating a location for object storage.

Creating a Request

data UpdateLocationObjectStorage

See: newUpdateLocationObjectStorage smart constructor.

Constructors

UpdateLocationObjectStorage'

Fields

serverPort :: Maybe Natural

The port that your self-managed object storage server accepts inbound network traffic on. The server port is set by default to TCP 80 (HTTP) or TCP 443 (HTTPS). You can specify a custom port if your self-managed object storage server requires one.

agentArns :: Maybe (NonEmpty Text)

The Amazon Resource Name (ARN) of the agents associated with the self-managed object storage server location.

serverProtocol :: Maybe ObjectStorageServerProtocol

The protocol that the object storage server uses to communicate. Valid values are HTTP or HTTPS.

secretKey :: Maybe (Sensitive Text)

Optional. The secret key is used if credentials are required to access the self-managed object storage server. If your object storage requires a user name and password to authenticate, use AccessKey and SecretKey to provide the user name and password, respectively.

subdirectory :: Maybe Text

The subdirectory in the self-managed object storage server that is used to read data from.

accessKey :: Maybe Text

Optional. The access key is used if credentials are required to access the self-managed object storage server. If your object storage requires a user name and password to authenticate, use AccessKey and SecretKey to provide the user name and password, respectively.

locationArn :: Text

The Amazon Resource Name (ARN) of the self-managed object storage server location to be updated.

newUpdateLocationObjectStorage

Arguments

:: Text
$sel:locationArn:UpdateLocationObjectStorage'

-> UpdateLocationObjectStorage

Create a value of UpdateLocationObjectStorage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serverPort:UpdateLocationObjectStorage', updateLocationObjectStorage_serverPort - The port that your self-managed object storage server accepts inbound network traffic on. The server port is set by default to TCP 80 (HTTP) or TCP 443 (HTTPS). You can specify a custom port if your self-managed object storage server requires one.

$sel:agentArns:UpdateLocationObjectStorage', updateLocationObjectStorage_agentArns - The Amazon Resource Name (ARN) of the agents associated with the self-managed object storage server location.

$sel:serverProtocol:UpdateLocationObjectStorage', updateLocationObjectStorage_serverProtocol - The protocol that the object storage server uses to communicate. Valid values are HTTP or HTTPS.

$sel:secretKey:UpdateLocationObjectStorage', updateLocationObjectStorage_secretKey - Optional. The secret key is used if credentials are required to access the self-managed object storage server. If your object storage requires a user name and password to authenticate, use AccessKey and SecretKey to provide the user name and password, respectively.

$sel:subdirectory:UpdateLocationObjectStorage', updateLocationObjectStorage_subdirectory - The subdirectory in the self-managed object storage server that is used to read data from.

$sel:accessKey:UpdateLocationObjectStorage', updateLocationObjectStorage_accessKey - Optional. The access key is used if credentials are required to access the self-managed object storage server. If your object storage requires a user name and password to authenticate, use AccessKey and SecretKey to provide the user name and password, respectively.

$sel:locationArn:UpdateLocationObjectStorage', updateLocationObjectStorage_locationArn - The Amazon Resource Name (ARN) of the self-managed object storage server location to be updated.

Request Lenses

updateLocationObjectStorage_serverPort :: Lens' UpdateLocationObjectStorage (Maybe Natural)

The port that your self-managed object storage server accepts inbound network traffic on. The server port is set by default to TCP 80 (HTTP) or TCP 443 (HTTPS). You can specify a custom port if your self-managed object storage server requires one.

updateLocationObjectStorage_agentArns :: Lens' UpdateLocationObjectStorage (Maybe (NonEmpty Text))

The Amazon Resource Name (ARN) of the agents associated with the self-managed object storage server location.

updateLocationObjectStorage_serverProtocol :: Lens' UpdateLocationObjectStorage (Maybe ObjectStorageServerProtocol)

The protocol that the object storage server uses to communicate. Valid values are HTTP or HTTPS.

updateLocationObjectStorage_secretKey :: Lens' UpdateLocationObjectStorage (Maybe Text)

Optional. The secret key is used if credentials are required to access the self-managed object storage server. If your object storage requires a user name and password to authenticate, use AccessKey and SecretKey to provide the user name and password, respectively.

updateLocationObjectStorage_subdirectory :: Lens' UpdateLocationObjectStorage (Maybe Text)

The subdirectory in the self-managed object storage server that is used to read data from.

updateLocationObjectStorage_accessKey :: Lens' UpdateLocationObjectStorage (Maybe Text)

Optional. The access key is used if credentials are required to access the self-managed object storage server. If your object storage requires a user name and password to authenticate, use AccessKey and SecretKey to provide the user name and password, respectively.

updateLocationObjectStorage_locationArn :: Lens' UpdateLocationObjectStorage Text

The Amazon Resource Name (ARN) of the self-managed object storage server location to be updated.

Destructuring the Response

data UpdateLocationObjectStorageResponse

See: newUpdateLocationObjectStorageResponse smart constructor.

Constructors

UpdateLocationObjectStorageResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateLocationObjectStorageResponse

Arguments

:: Int
$sel:httpStatus:UpdateLocationObjectStorageResponse'

-> UpdateLocationObjectStorageResponse

Create a value of UpdateLocationObjectStorageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateLocationObjectStorageResponse', updateLocationObjectStorageResponse_httpStatus - The response's http status code.

Response Lenses

updateLocationObjectStorageResponse_httpStatus :: Lens' UpdateLocationObjectStorageResponse Int

The response's http status code.