Network.AWS.DataSync.UpdateLocationSmb

Description

Updates some of the parameters of a previously created location for Server Message Block (SMB) file system access. For information about creating an SMB location, see Creating a location for SMB.

Creating a Request

data UpdateLocationSmb

See: newUpdateLocationSmb smart constructor.

Constructors

UpdateLocationSmb'

Fields

agentArns :: Maybe (NonEmpty Text)

The Amazon Resource Names (ARNs) of agents to use for a Simple Message Block (SMB) location.

domain :: Maybe Text

The name of the Windows domain that the SMB server belongs to.

user :: Maybe Text

The user who can mount the share has the permissions to access files and folders in the SMB share.

password :: Maybe (Sensitive Text)

The password of the user who can mount the share has the permissions to access files and folders in the SMB share.

subdirectory :: Maybe Text

The subdirectory in the SMB file system that is used to read data from the SMB source location or write data to the SMB destination. The SMB path should be a path that's exported by the SMB server, or a subdirectory of that path. The path should be such that it can be mounted by other SMB clients in your network.

Subdirectory must be specified with forward slashes. For example, /path/to/folder.

To transfer all the data in the folder that you specified, DataSync must have permissions to mount the SMB share and to access all the data in that share. To ensure this, do either of the following:

Ensure that the user/password specified belongs to the user who can mount the share and who has the appropriate permissions for all of the files and directories that you want DataSync to access.

Use credentials of a member of the Backup Operators group to mount the share.

Doing either of these options enables the agent to access the data. For the agent to access directories, you must also enable all execute access.

mountOptions :: Maybe SmbMountOptions

locationArn :: Text

The Amazon Resource Name (ARN) of the SMB location to update.

newUpdateLocationSmb

Arguments

:: Text
$sel:locationArn:UpdateLocationSmb'

-> UpdateLocationSmb

Create a value of UpdateLocationSmb with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:agentArns:UpdateLocationSmb', updateLocationSmb_agentArns - The Amazon Resource Names (ARNs) of agents to use for a Simple Message Block (SMB) location.

$sel:domain:UpdateLocationSmb', updateLocationSmb_domain - The name of the Windows domain that the SMB server belongs to.

$sel:user:UpdateLocationSmb', updateLocationSmb_user - The user who can mount the share has the permissions to access files and folders in the SMB share.

$sel:password:UpdateLocationSmb', updateLocationSmb_password - The password of the user who can mount the share has the permissions to access files and folders in the SMB share.

$sel:subdirectory:UpdateLocationSmb', updateLocationSmb_subdirectory - The subdirectory in the SMB file system that is used to read data from the SMB source location or write data to the SMB destination. The SMB path should be a path that's exported by the SMB server, or a subdirectory of that path. The path should be such that it can be mounted by other SMB clients in your network.

Subdirectory must be specified with forward slashes. For example, /path/to/folder.

To transfer all the data in the folder that you specified, DataSync must have permissions to mount the SMB share and to access all the data in that share. To ensure this, do either of the following:

Ensure that the user/password specified belongs to the user who can mount the share and who has the appropriate permissions for all of the files and directories that you want DataSync to access.

Use credentials of a member of the Backup Operators group to mount the share.

Doing either of these options enables the agent to access the data. For the agent to access directories, you must also enable all execute access.

$sel:mountOptions:UpdateLocationSmb', updateLocationSmb_mountOptions - Undocumented member.

$sel:locationArn:UpdateLocationSmb', updateLocationSmb_locationArn - The Amazon Resource Name (ARN) of the SMB location to update.

Request Lenses

updateLocationSmb_agentArns :: Lens' UpdateLocationSmb (Maybe (NonEmpty Text))

The Amazon Resource Names (ARNs) of agents to use for a Simple Message Block (SMB) location.

updateLocationSmb_domain :: Lens' UpdateLocationSmb (Maybe Text)

The name of the Windows domain that the SMB server belongs to.

updateLocationSmb_user :: Lens' UpdateLocationSmb (Maybe Text)

The user who can mount the share has the permissions to access files and folders in the SMB share.

updateLocationSmb_password :: Lens' UpdateLocationSmb (Maybe Text)

The password of the user who can mount the share has the permissions to access files and folders in the SMB share.

updateLocationSmb_subdirectory :: Lens' UpdateLocationSmb (Maybe Text)

The subdirectory in the SMB file system that is used to read data from the SMB source location or write data to the SMB destination. The SMB path should be a path that's exported by the SMB server, or a subdirectory of that path. The path should be such that it can be mounted by other SMB clients in your network.

Subdirectory must be specified with forward slashes. For example, /path/to/folder.

To transfer all the data in the folder that you specified, DataSync must have permissions to mount the SMB share and to access all the data in that share. To ensure this, do either of the following:

Ensure that the user/password specified belongs to the user who can mount the share and who has the appropriate permissions for all of the files and directories that you want DataSync to access.

Use credentials of a member of the Backup Operators group to mount the share.

Doing either of these options enables the agent to access the data. For the agent to access directories, you must also enable all execute access.

updateLocationSmb_mountOptions :: Lens' UpdateLocationSmb (Maybe SmbMountOptions)

Undocumented member.

updateLocationSmb_locationArn :: Lens' UpdateLocationSmb Text

The Amazon Resource Name (ARN) of the SMB location to update.

Destructuring the Response

data UpdateLocationSmbResponse

See: newUpdateLocationSmbResponse smart constructor.

Constructors

UpdateLocationSmbResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateLocationSmbResponse

Arguments

:: Int
$sel:httpStatus:UpdateLocationSmbResponse'

-> UpdateLocationSmbResponse

Create a value of UpdateLocationSmbResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateLocationSmbResponse', updateLocationSmbResponse_httpStatus - The response's http status code.

Response Lenses

updateLocationSmbResponse_httpStatus :: Lens' UpdateLocationSmbResponse Int

The response's http status code.