Network.AWS.Detective.CreateMembers

Description

Sends a request to invite the specified AWS accounts to be member accounts in the behavior graph. This operation can only be called by the administrator account for a behavior graph.

CreateMembers verifies the accounts and then invites the verified accounts. The administrator can optionally specify to not send invitation emails to the member accounts. This would be used when the administrator manages their member accounts centrally.

The request provides the behavior graph ARN and the list of accounts to invite.

The response separates the requested accounts into two lists:

The accounts that CreateMembers was able to start the verification for. This list includes member accounts that are being verified, that have passed verification and are to be invited, and that have failed verification.

The accounts that CreateMembers was unable to process. This list includes accounts that were already invited to be member accounts in the behavior graph.

Creating a Request

data CreateMembers

See: newCreateMembers smart constructor.

Constructors

CreateMembers'

Fields

disableEmailNotification :: Maybe Bool

if set to true, then the member accounts do not receive email notifications. By default, this is set to false, and the member accounts receive email notifications.

message :: Maybe Text

Customized message text to include in the invitation email message to the invited member accounts.

graphArn :: Text

The ARN of the behavior graph to invite the member accounts to contribute their data to.

accounts :: NonEmpty Account

The list of AWS accounts to invite to become member accounts in the behavior graph. You can invite up to 50 accounts at a time. For each invited account, the account list contains the account identifier and the AWS account root user email address.

newCreateMembers

Arguments

:: Text
$sel:graphArn:CreateMembers'

-> NonEmpty Account
$sel:accounts:CreateMembers'

-> CreateMembers

Create a value of CreateMembers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:disableEmailNotification:CreateMembers', createMembers_disableEmailNotification - if set to true, then the member accounts do not receive email notifications. By default, this is set to false, and the member accounts receive email notifications.

$sel:message:CreateMembers', createMembers_message - Customized message text to include in the invitation email message to the invited member accounts.

$sel:graphArn:CreateMembers', createMembers_graphArn - The ARN of the behavior graph to invite the member accounts to contribute their data to.

$sel:accounts:CreateMembers', createMembers_accounts - The list of AWS accounts to invite to become member accounts in the behavior graph. You can invite up to 50 accounts at a time. For each invited account, the account list contains the account identifier and the AWS account root user email address.

Request Lenses

createMembers_disableEmailNotification :: Lens' CreateMembers (Maybe Bool)

if set to true, then the member accounts do not receive email notifications. By default, this is set to false, and the member accounts receive email notifications.

createMembers_message :: Lens' CreateMembers (Maybe Text)

Customized message text to include in the invitation email message to the invited member accounts.

createMembers_graphArn :: Lens' CreateMembers Text

The ARN of the behavior graph to invite the member accounts to contribute their data to.

createMembers_accounts :: Lens' CreateMembers (NonEmpty Account)

The list of AWS accounts to invite to become member accounts in the behavior graph. You can invite up to 50 accounts at a time. For each invited account, the account list contains the account identifier and the AWS account root user email address.

Destructuring the Response

data CreateMembersResponse

See: newCreateMembersResponse smart constructor.

Constructors

CreateMembersResponse'

Fields

members :: Maybe [MemberDetail]

The set of member account invitation requests that Detective was able to process. This includes accounts that are being verified, that failed verification, and that passed verification and are being sent an invitation.

unprocessedAccounts :: Maybe [UnprocessedAccount]

The list of accounts for which Detective was unable to process the invitation request. For each account, the list provides the reason why the request could not be processed. The list includes accounts that are already member accounts in the behavior graph.

httpStatus :: Int

The response's http status code.

newCreateMembersResponse

Arguments

:: Int
$sel:httpStatus:CreateMembersResponse'

-> CreateMembersResponse

Create a value of CreateMembersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:members:CreateMembersResponse', createMembersResponse_members - The set of member account invitation requests that Detective was able to process. This includes accounts that are being verified, that failed verification, and that passed verification and are being sent an invitation.

$sel:unprocessedAccounts:CreateMembersResponse', createMembersResponse_unprocessedAccounts - The list of accounts for which Detective was unable to process the invitation request. For each account, the list provides the reason why the request could not be processed. The list includes accounts that are already member accounts in the behavior graph.

$sel:httpStatus:CreateMembersResponse', createMembersResponse_httpStatus - The response's http status code.

Response Lenses

createMembersResponse_members :: Lens' CreateMembersResponse (Maybe [MemberDetail])

The set of member account invitation requests that Detective was able to process. This includes accounts that are being verified, that failed verification, and that passed verification and are being sent an invitation.

createMembersResponse_unprocessedAccounts :: Lens' CreateMembersResponse (Maybe [UnprocessedAccount])

The list of accounts for which Detective was unable to process the invitation request. For each account, the list provides the reason why the request could not be processed. The list includes accounts that are already member accounts in the behavior graph.

createMembersResponse_httpStatus :: Lens' CreateMembersResponse Int

The response's http status code.