Network.AWS.Detective.ListInvitations

Description

Retrieves the list of open and accepted behavior graph invitations for the member account. This operation can only be called by a member account.

Open invitations are invitations that the member account has not responded to.

The results do not include behavior graphs for which the member account declined the invitation. The results also do not include behavior graphs that the member account resigned from or was removed from.

Creating a Request

data ListInvitations

See: newListInvitations smart constructor.

Constructors

ListInvitations'

Fields

nextToken :: Maybe Text

For requests to retrieve the next page of results, the pagination token that was returned with the previous page of results. The initial request does not include a pagination token.

maxResults :: Maybe Natural

The maximum number of behavior graph invitations to return in the response. The total must be less than the overall limit on the number of results to return, which is currently 200.

newListInvitations :: ListInvitations

Create a value of ListInvitations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListInvitations', listInvitations_nextToken - For requests to retrieve the next page of results, the pagination token that was returned with the previous page of results. The initial request does not include a pagination token.

$sel:maxResults:ListInvitations', listInvitations_maxResults - The maximum number of behavior graph invitations to return in the response. The total must be less than the overall limit on the number of results to return, which is currently 200.

Request Lenses

listInvitations_nextToken :: Lens' ListInvitations (Maybe Text)

For requests to retrieve the next page of results, the pagination token that was returned with the previous page of results. The initial request does not include a pagination token.

listInvitations_maxResults :: Lens' ListInvitations (Maybe Natural)

The maximum number of behavior graph invitations to return in the response. The total must be less than the overall limit on the number of results to return, which is currently 200.

Destructuring the Response

data ListInvitationsResponse

See: newListInvitationsResponse smart constructor.

Constructors

ListInvitationsResponse'

Fields

invitations :: Maybe [MemberDetail]

The list of behavior graphs for which the member account has open or accepted invitations.

nextToken :: Maybe Text

If there are more behavior graphs remaining in the results, then this is the pagination token to use to request the next page of behavior graphs.

httpStatus :: Int

The response's http status code.

newListInvitationsResponse

Arguments

:: Int
$sel:httpStatus:ListInvitationsResponse'

-> ListInvitationsResponse

Create a value of ListInvitationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:invitations:ListInvitationsResponse', listInvitationsResponse_invitations - The list of behavior graphs for which the member account has open or accepted invitations.

$sel:nextToken:ListInvitations', listInvitationsResponse_nextToken - If there are more behavior graphs remaining in the results, then this is the pagination token to use to request the next page of behavior graphs.

$sel:httpStatus:ListInvitationsResponse', listInvitationsResponse_httpStatus - The response's http status code.

Response Lenses

listInvitationsResponse_invitations :: Lens' ListInvitationsResponse (Maybe [MemberDetail])

The list of behavior graphs for which the member account has open or accepted invitations.

listInvitationsResponse_nextToken :: Lens' ListInvitationsResponse (Maybe Text)

If there are more behavior graphs remaining in the results, then this is the pagination token to use to request the next page of behavior graphs.

listInvitationsResponse_httpStatus :: Lens' ListInvitationsResponse Int

The response's http status code.