Network.AWS.Detective.ListMembers

Description

Retrieves the list of member accounts for a behavior graph. Does not return member accounts that were removed from the behavior graph.

Creating a Request

data ListMembers

See: newListMembers smart constructor.

Constructors

ListMembers'

Fields

nextToken :: Maybe Text

For requests to retrieve the next page of member account results, the pagination token that was returned with the previous page of results. The initial request does not include a pagination token.

maxResults :: Maybe Natural

The maximum number of member accounts to include in the response. The total must be less than the overall limit on the number of results to return, which is currently 200.

graphArn :: Text

The ARN of the behavior graph for which to retrieve the list of member accounts.

newListMembers

Arguments

:: Text
$sel:graphArn:ListMembers'

-> ListMembers

Create a value of ListMembers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListMembers', listMembers_nextToken - For requests to retrieve the next page of member account results, the pagination token that was returned with the previous page of results. The initial request does not include a pagination token.

$sel:maxResults:ListMembers', listMembers_maxResults - The maximum number of member accounts to include in the response. The total must be less than the overall limit on the number of results to return, which is currently 200.

$sel:graphArn:ListMembers', listMembers_graphArn - The ARN of the behavior graph for which to retrieve the list of member accounts.

Request Lenses

listMembers_nextToken :: Lens' ListMembers (Maybe Text)

For requests to retrieve the next page of member account results, the pagination token that was returned with the previous page of results. The initial request does not include a pagination token.

listMembers_maxResults :: Lens' ListMembers (Maybe Natural)

The maximum number of member accounts to include in the response. The total must be less than the overall limit on the number of results to return, which is currently 200.

listMembers_graphArn :: Lens' ListMembers Text

The ARN of the behavior graph for which to retrieve the list of member accounts.

Destructuring the Response

data ListMembersResponse

See: newListMembersResponse smart constructor.

Constructors

ListMembersResponse'

Fields

memberDetails :: Maybe [MemberDetail]

The list of member accounts in the behavior graph.

The results include member accounts that did not pass verification and member accounts that have not yet accepted the invitation to the behavior graph. The results do not include member accounts that were removed from the behavior graph.

nextToken :: Maybe Text

If there are more member accounts remaining in the results, then this is the pagination token to use to request the next page of member accounts.

httpStatus :: Int

The response's http status code.

newListMembersResponse

Arguments

:: Int
$sel:httpStatus:ListMembersResponse'

-> ListMembersResponse

Create a value of ListMembersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:memberDetails:ListMembersResponse', listMembersResponse_memberDetails - The list of member accounts in the behavior graph.

The results include member accounts that did not pass verification and member accounts that have not yet accepted the invitation to the behavior graph. The results do not include member accounts that were removed from the behavior graph.

$sel:nextToken:ListMembers', listMembersResponse_nextToken - If there are more member accounts remaining in the results, then this is the pagination token to use to request the next page of member accounts.

$sel:httpStatus:ListMembersResponse', listMembersResponse_httpStatus - The response's http status code.

Response Lenses

listMembersResponse_memberDetails :: Lens' ListMembersResponse (Maybe [MemberDetail])

The list of member accounts in the behavior graph.

The results include member accounts that did not pass verification and member accounts that have not yet accepted the invitation to the behavior graph. The results do not include member accounts that were removed from the behavior graph.

listMembersResponse_nextToken :: Lens' ListMembersResponse (Maybe Text)

If there are more member accounts remaining in the results, then this is the pagination token to use to request the next page of member accounts.

listMembersResponse_httpStatus :: Lens' ListMembersResponse Int

The response's http status code.