Network.AWS.Detective.Types.MemberDetail

Description

Documentation

data MemberDetail

Details about a member account that was invited to contribute to a behavior graph.

See: newMemberDetail smart constructor.

Constructors

MemberDetail'

Fields

percentOfGraphUtilizationUpdatedTime :: Maybe POSIX

The date and time when the graph utilization percentage was last updated.

status :: Maybe MemberStatus

The current membership status of the member account. The status can have one of the following values:

INVITED - Indicates that the member was sent an invitation but has not yet responded.

VERIFICATION_IN_PROGRESS - Indicates that Detective is verifying that the account identifier and email address provided for the member account match. If they do match, then Detective sends the invitation. If the email address and account identifier don't match, then the member cannot be added to the behavior graph.

VERIFICATION_FAILED - Indicates that the account and email address provided for the member account do not match, and Detective did not send an invitation to the account.

ENABLED - Indicates that the member account accepted the invitation to contribute to the behavior graph.

ACCEPTED_BUT_DISABLED - Indicates that the member account accepted the invitation but is prevented from contributing data to the behavior graph. DisabledReason provides the reason why the member account is not enabled.

Member accounts that declined an invitation or that were removed from the behavior graph are not included.

invitedTime :: Maybe POSIX

The date and time that Detective sent the invitation to the member account. The value is in milliseconds since the epoch.

administratorId :: Maybe Text

The AWS account identifier of the administrator account for the behavior graph.

graphArn :: Maybe Text

The ARN of the behavior graph that the member account was invited to.

masterId :: Maybe Text

The AWS account identifier of the administrator account for the behavior graph.

accountId :: Maybe Text

The AWS account identifier for the member account.

disabledReason :: Maybe MemberDisabledReason

For member accounts with a status of ACCEPTED_BUT_DISABLED, the reason that the member account is not enabled.

The reason can have one of the following values:

VOLUME_TOO_HIGH - Indicates that adding the member account would cause the data volume for the behavior graph to be too high.

VOLUME_UNKNOWN - Indicates that Detective is unable to verify the data volume for the member account. This is usually because the member account is not enrolled in Amazon GuardDuty.

percentOfGraphUtilization :: Maybe Double

The member account data volume as a percentage of the maximum allowed data volume. 0 indicates 0 percent, and 100 indicates 100 percent.

Note that this is not the percentage of the behavior graph data volume.

For example, the data volume for the behavior graph is 80 GB per day. The maximum data volume is 160 GB per day. If the data volume for the member account is 40 GB per day, then PercentOfGraphUtilization is 25. It represents 25% of the maximum allowed data volume.

emailAddress :: Maybe Text

The AWS account root user email address for the member account.

volumeUsageUpdatedTime :: Maybe POSIX

The data and time when the member account data volume was last updated.

updatedTime :: Maybe POSIX

The date and time that the member account was last updated. The value is in milliseconds since the epoch.

volumeUsageInBytes :: Maybe Integer

The data volume in bytes per day for the member account.

newMemberDetail :: MemberDetail

Create a value of MemberDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:percentOfGraphUtilizationUpdatedTime:MemberDetail', memberDetail_percentOfGraphUtilizationUpdatedTime - The date and time when the graph utilization percentage was last updated.

$sel:status:MemberDetail', memberDetail_status - The current membership status of the member account. The status can have one of the following values:

INVITED - Indicates that the member was sent an invitation but has not yet responded.

VERIFICATION_IN_PROGRESS - Indicates that Detective is verifying that the account identifier and email address provided for the member account match. If they do match, then Detective sends the invitation. If the email address and account identifier don't match, then the member cannot be added to the behavior graph.

VERIFICATION_FAILED - Indicates that the account and email address provided for the member account do not match, and Detective did not send an invitation to the account.

ENABLED - Indicates that the member account accepted the invitation to contribute to the behavior graph.

ACCEPTED_BUT_DISABLED - Indicates that the member account accepted the invitation but is prevented from contributing data to the behavior graph. DisabledReason provides the reason why the member account is not enabled.

Member accounts that declined an invitation or that were removed from the behavior graph are not included.

$sel:invitedTime:MemberDetail', memberDetail_invitedTime - The date and time that Detective sent the invitation to the member account. The value is in milliseconds since the epoch.

$sel:administratorId:MemberDetail', memberDetail_administratorId - The AWS account identifier of the administrator account for the behavior graph.

$sel:graphArn:MemberDetail', memberDetail_graphArn - The ARN of the behavior graph that the member account was invited to.

$sel:masterId:MemberDetail', memberDetail_masterId - The AWS account identifier of the administrator account for the behavior graph.

$sel:accountId:MemberDetail', memberDetail_accountId - The AWS account identifier for the member account.

$sel:disabledReason:MemberDetail', memberDetail_disabledReason - For member accounts with a status of ACCEPTED_BUT_DISABLED, the reason that the member account is not enabled.

The reason can have one of the following values:

VOLUME_TOO_HIGH - Indicates that adding the member account would cause the data volume for the behavior graph to be too high.

VOLUME_UNKNOWN - Indicates that Detective is unable to verify the data volume for the member account. This is usually because the member account is not enrolled in Amazon GuardDuty.

$sel:percentOfGraphUtilization:MemberDetail', memberDetail_percentOfGraphUtilization - The member account data volume as a percentage of the maximum allowed data volume. 0 indicates 0 percent, and 100 indicates 100 percent.

Note that this is not the percentage of the behavior graph data volume.

For example, the data volume for the behavior graph is 80 GB per day. The maximum data volume is 160 GB per day. If the data volume for the member account is 40 GB per day, then PercentOfGraphUtilization is 25. It represents 25% of the maximum allowed data volume.

$sel:emailAddress:MemberDetail', memberDetail_emailAddress - The AWS account root user email address for the member account.

$sel:volumeUsageUpdatedTime:MemberDetail', memberDetail_volumeUsageUpdatedTime - The data and time when the member account data volume was last updated.

$sel:updatedTime:MemberDetail', memberDetail_updatedTime - The date and time that the member account was last updated. The value is in milliseconds since the epoch.

$sel:volumeUsageInBytes:MemberDetail', memberDetail_volumeUsageInBytes - The data volume in bytes per day for the member account.

memberDetail_percentOfGraphUtilizationUpdatedTime :: Lens' MemberDetail (Maybe UTCTime)

The date and time when the graph utilization percentage was last updated.

memberDetail_status :: Lens' MemberDetail (Maybe MemberStatus)

The current membership status of the member account. The status can have one of the following values:

INVITED - Indicates that the member was sent an invitation but has not yet responded.

VERIFICATION_IN_PROGRESS - Indicates that Detective is verifying that the account identifier and email address provided for the member account match. If they do match, then Detective sends the invitation. If the email address and account identifier don't match, then the member cannot be added to the behavior graph.

VERIFICATION_FAILED - Indicates that the account and email address provided for the member account do not match, and Detective did not send an invitation to the account.

ENABLED - Indicates that the member account accepted the invitation to contribute to the behavior graph.

ACCEPTED_BUT_DISABLED - Indicates that the member account accepted the invitation but is prevented from contributing data to the behavior graph. DisabledReason provides the reason why the member account is not enabled.

Member accounts that declined an invitation or that were removed from the behavior graph are not included.

memberDetail_invitedTime :: Lens' MemberDetail (Maybe UTCTime)

The date and time that Detective sent the invitation to the member account. The value is in milliseconds since the epoch.

memberDetail_administratorId :: Lens' MemberDetail (Maybe Text)

The AWS account identifier of the administrator account for the behavior graph.

memberDetail_graphArn :: Lens' MemberDetail (Maybe Text)

The ARN of the behavior graph that the member account was invited to.

memberDetail_masterId :: Lens' MemberDetail (Maybe Text)

The AWS account identifier of the administrator account for the behavior graph.

memberDetail_accountId :: Lens' MemberDetail (Maybe Text)

The AWS account identifier for the member account.

memberDetail_disabledReason :: Lens' MemberDetail (Maybe MemberDisabledReason)

For member accounts with a status of ACCEPTED_BUT_DISABLED, the reason that the member account is not enabled.

The reason can have one of the following values:

VOLUME_TOO_HIGH - Indicates that adding the member account would cause the data volume for the behavior graph to be too high.

VOLUME_UNKNOWN - Indicates that Detective is unable to verify the data volume for the member account. This is usually because the member account is not enrolled in Amazon GuardDuty.

memberDetail_percentOfGraphUtilization :: Lens' MemberDetail (Maybe Double)

The member account data volume as a percentage of the maximum allowed data volume. 0 indicates 0 percent, and 100 indicates 100 percent.

Note that this is not the percentage of the behavior graph data volume.

For example, the data volume for the behavior graph is 80 GB per day. The maximum data volume is 160 GB per day. If the data volume for the member account is 40 GB per day, then PercentOfGraphUtilization is 25. It represents 25% of the maximum allowed data volume.

memberDetail_emailAddress :: Lens' MemberDetail (Maybe Text)

The AWS account root user email address for the member account.

memberDetail_volumeUsageUpdatedTime :: Lens' MemberDetail (Maybe UTCTime)

The data and time when the member account data volume was last updated.

memberDetail_updatedTime :: Lens' MemberDetail (Maybe UTCTime)

The date and time that the member account was last updated. The value is in milliseconds since the epoch.

memberDetail_volumeUsageInBytes :: Lens' MemberDetail (Maybe Integer)

The data volume in bytes per day for the member account.