Network.AWS.DevOpsGuru.DescribeAnomaly

Description

Returns details about an anomaly that you specify using its ID.

Creating a Request

data DescribeAnomaly

See: newDescribeAnomaly smart constructor.

Constructors

DescribeAnomaly'

Fields

id :: Text

The ID of the anomaly.

newDescribeAnomaly

Arguments

:: Text
$sel:id:DescribeAnomaly'

-> DescribeAnomaly

Create a value of DescribeAnomaly with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:DescribeAnomaly', describeAnomaly_id - The ID of the anomaly.

Request Lenses

describeAnomaly_id :: Lens' DescribeAnomaly Text

The ID of the anomaly.

Destructuring the Response

data DescribeAnomalyResponse

See: newDescribeAnomalyResponse smart constructor.

Constructors

DescribeAnomalyResponse'

Fields

reactiveAnomaly :: Maybe ReactiveAnomaly

A ProactiveAnomaly object that represents the requested anomaly.

proactiveAnomaly :: Maybe ProactiveAnomaly

A ReactiveAnomaly object that represents the requested anomaly.

httpStatus :: Int

The response's http status code.

newDescribeAnomalyResponse

Arguments

:: Int
$sel:httpStatus:DescribeAnomalyResponse'

-> DescribeAnomalyResponse

Create a value of DescribeAnomalyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reactiveAnomaly:DescribeAnomalyResponse', describeAnomalyResponse_reactiveAnomaly - A ProactiveAnomaly object that represents the requested anomaly.

$sel:proactiveAnomaly:DescribeAnomalyResponse', describeAnomalyResponse_proactiveAnomaly - A ReactiveAnomaly object that represents the requested anomaly.

$sel:httpStatus:DescribeAnomalyResponse', describeAnomalyResponse_httpStatus - The response's http status code.

Response Lenses

describeAnomalyResponse_reactiveAnomaly :: Lens' DescribeAnomalyResponse (Maybe ReactiveAnomaly)

A ProactiveAnomaly object that represents the requested anomaly.

describeAnomalyResponse_proactiveAnomaly :: Lens' DescribeAnomalyResponse (Maybe ProactiveAnomaly)

A ReactiveAnomaly object that represents the requested anomaly.

describeAnomalyResponse_httpStatus :: Lens' DescribeAnomalyResponse Int

The response's http status code.