Network.AWS.DevOpsGuru.Lens

Description

Operations

DescribeAnomaly

describeAnomaly_id :: Lens' DescribeAnomaly Text

The ID of the anomaly.

describeAnomalyResponse_reactiveAnomaly :: Lens' DescribeAnomalyResponse (Maybe ReactiveAnomaly)

A ProactiveAnomaly object that represents the requested anomaly.

describeAnomalyResponse_proactiveAnomaly :: Lens' DescribeAnomalyResponse (Maybe ProactiveAnomaly)

A ReactiveAnomaly object that represents the requested anomaly.

describeAnomalyResponse_httpStatus :: Lens' DescribeAnomalyResponse Int

The response's http status code.

DescribeFeedback

describeFeedback_insightId :: Lens' DescribeFeedback (Maybe Text)

The ID of the insight for which the feedback was provided.

describeFeedbackResponse_insightFeedback :: Lens' DescribeFeedbackResponse (Maybe InsightFeedback)

Undocumented member.

describeFeedbackResponse_httpStatus :: Lens' DescribeFeedbackResponse Int

The response's http status code.

ListInsights

listInsights_nextToken :: Lens' ListInsights (Maybe Text)

The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

listInsights_maxResults :: Lens' ListInsights (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

listInsights_statusFilter :: Lens' ListInsights ListInsightsStatusFilter

A filter used to filter the returned insights by their status. You can specify one status filter.

listInsightsResponse_reactiveInsights :: Lens' ListInsightsResponse (Maybe [ReactiveInsightSummary])

The returned list of reactive insights.

listInsightsResponse_nextToken :: Lens' ListInsightsResponse (Maybe Text)

The pagination token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

listInsightsResponse_proactiveInsights :: Lens' ListInsightsResponse (Maybe [ProactiveInsightSummary])

The returned list of proactive insights.

listInsightsResponse_httpStatus :: Lens' ListInsightsResponse Int

The response's http status code.

AddNotificationChannel

addNotificationChannel_config :: Lens' AddNotificationChannel NotificationChannelConfig

A NotificationChannelConfig object that specifies what type of notification channel to add. The one supported notification channel is Amazon Simple Notification Service (Amazon SNS).

addNotificationChannelResponse_httpStatus :: Lens' AddNotificationChannelResponse Int

The response's http status code.

addNotificationChannelResponse_id :: Lens' AddNotificationChannelResponse Text

The ID of the added notification channel.

ListNotificationChannels

listNotificationChannels_nextToken :: Lens' ListNotificationChannels (Maybe Text)

The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

listNotificationChannelsResponse_channels :: Lens' ListNotificationChannelsResponse (Maybe [NotificationChannel])

An array that contains the requested notification channels.

listNotificationChannelsResponse_nextToken :: Lens' ListNotificationChannelsResponse (Maybe Text)

The pagination token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

listNotificationChannelsResponse_httpStatus :: Lens' ListNotificationChannelsResponse Int

The response's http status code.

DescribeAccountOverview

describeAccountOverview_toTime :: Lens' DescribeAccountOverview (Maybe UTCTime)

The end of the time range passed in. The start time granularity is at the day level. The floor of the start time is used. Returned information occurred before this day. If this is not specified, then the current day is used.

describeAccountOverview_fromTime :: Lens' DescribeAccountOverview UTCTime

The start of the time range passed in. The start time granularity is at the day level. The floor of the start time is used. Returned information occurred after this day.

describeAccountOverviewResponse_httpStatus :: Lens' DescribeAccountOverviewResponse Int

The response's http status code.

describeAccountOverviewResponse_reactiveInsights :: Lens' DescribeAccountOverviewResponse Int

An integer that specifies the number of open reactive insights in your AWS account that were created during the time range passed in.

describeAccountOverviewResponse_proactiveInsights :: Lens' DescribeAccountOverviewResponse Int

An integer that specifies the number of open proactive insights in your AWS account that were created during the time range passed in.

describeAccountOverviewResponse_meanTimeToRecoverInMilliseconds :: Lens' DescribeAccountOverviewResponse Integer

The Mean Time to Recover (MTTR) for all closed insights that were created during the time range passed in.

DescribeResourceCollectionHealth

describeResourceCollectionHealth_nextToken :: Lens' DescribeResourceCollectionHealth (Maybe Text)

The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

describeResourceCollectionHealth_resourceCollectionType :: Lens' DescribeResourceCollectionHealth ResourceCollectionType

An AWS resource collection type. This type specifies how analyzed AWS resources are defined. The one type of AWS resource collection supported is AWS CloudFormation stacks. DevOps Guru can be configured to analyze only the AWS resources that are defined in the stacks. You can specify up to 500 AWS CloudFormation stacks.

describeResourceCollectionHealthResponse_service :: Lens' DescribeResourceCollectionHealthResponse (Maybe [ServiceHealth])

An array of ServiceHealth objects that describes the health of the AWS services associated with the resources in the collection.

describeResourceCollectionHealthResponse_nextToken :: Lens' DescribeResourceCollectionHealthResponse (Maybe Text)

The pagination token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

describeResourceCollectionHealthResponse_httpStatus :: Lens' DescribeResourceCollectionHealthResponse Int

The response's http status code.

describeResourceCollectionHealthResponse_cloudFormation :: Lens' DescribeResourceCollectionHealthResponse [CloudFormationHealth]

The returned CloudFormationHealthOverview object that contains an InsightHealthOverview object with the requested system health information.

RemoveNotificationChannel

removeNotificationChannel_id :: Lens' RemoveNotificationChannel Text

The ID of the notification channel to be removed.

removeNotificationChannelResponse_httpStatus :: Lens' RemoveNotificationChannelResponse Int

The response's http status code.

ListAnomaliesForInsight

listAnomaliesForInsight_startTimeRange :: Lens' ListAnomaliesForInsight (Maybe StartTimeRange)

A time range used to specify when the requested anomalies started. All returned anomalies started during this time range.

listAnomaliesForInsight_nextToken :: Lens' ListAnomaliesForInsight (Maybe Text)

The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

listAnomaliesForInsight_maxResults :: Lens' ListAnomaliesForInsight (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

listAnomaliesForInsight_insightId :: Lens' ListAnomaliesForInsight Text

The ID of the insight. The returned anomalies belong to this insight.

listAnomaliesForInsightResponse_proactiveAnomalies :: Lens' ListAnomaliesForInsightResponse (Maybe [ProactiveAnomalySummary])

An array of ProactiveAnomalySummary objects that represent the requested anomalies

listAnomaliesForInsightResponse_nextToken :: Lens' ListAnomaliesForInsightResponse (Maybe Text)

The pagination token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

listAnomaliesForInsightResponse_reactiveAnomalies :: Lens' ListAnomaliesForInsightResponse (Maybe [ReactiveAnomalySummary])

An array of ReactiveAnomalySummary objects that represent the requested anomalies

listAnomaliesForInsightResponse_httpStatus :: Lens' ListAnomaliesForInsightResponse Int

The response's http status code.

PutFeedback

putFeedback_insightFeedback :: Lens' PutFeedback (Maybe InsightFeedback)

The feedback from customers is about the recommendations in this insight.

putFeedbackResponse_httpStatus :: Lens' PutFeedbackResponse Int

The response's http status code.

SearchInsights

searchInsights_filters :: Lens' SearchInsights (Maybe SearchInsightsFilters)

A SearchInsightsFilters object that is used to set the severity and status filters on your insight search.

searchInsights_nextToken :: Lens' SearchInsights (Maybe Text)

The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

searchInsights_maxResults :: Lens' SearchInsights (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

searchInsights_startTimeRange :: Lens' SearchInsights StartTimeRange

The start of the time range passed in. Returned insights occurred after this time.

searchInsights_type :: Lens' SearchInsights InsightType

The type of insights you are searching for (REACTIVE or PROACTIVE).

searchInsightsResponse_reactiveInsights :: Lens' SearchInsightsResponse (Maybe [ReactiveInsightSummary])

The returned reactive insights.

searchInsightsResponse_nextToken :: Lens' SearchInsightsResponse (Maybe Text)

The pagination token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

searchInsightsResponse_proactiveInsights :: Lens' SearchInsightsResponse (Maybe [ProactiveInsightSummary])

The returned proactive insights.

searchInsightsResponse_httpStatus :: Lens' SearchInsightsResponse Int

The response's http status code.

DescribeServiceIntegration

describeServiceIntegrationResponse_serviceIntegration :: Lens' DescribeServiceIntegrationResponse (Maybe ServiceIntegrationConfig)

Undocumented member.

describeServiceIntegrationResponse_httpStatus :: Lens' DescribeServiceIntegrationResponse Int

The response's http status code.

UpdateServiceIntegration

updateServiceIntegration_serviceIntegration :: Lens' UpdateServiceIntegration UpdateServiceIntegrationConfig

An IntegratedServiceConfig object used to specify the integrated service you want to update, and whether you want to update it to enabled or disabled.

updateServiceIntegrationResponse_httpStatus :: Lens' UpdateServiceIntegrationResponse Int

The response's http status code.

GetResourceCollection

getResourceCollection_nextToken :: Lens' GetResourceCollection (Maybe Text)

The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

getResourceCollection_resourceCollectionType :: Lens' GetResourceCollection ResourceCollectionType

The type of AWS resource collections to return. The one valid value is CLOUD_FORMATION for AWS CloudFormation stacks.

getResourceCollectionResponse_resourceCollection :: Lens' GetResourceCollectionResponse (Maybe ResourceCollectionFilter)

The requested list of AWS resource collections. The one type of AWS resource collection supported is AWS CloudFormation stacks. DevOps Guru can be configured to analyze only the AWS resources that are defined in the stacks. You can specify up to 500 AWS CloudFormation stacks.

getResourceCollectionResponse_nextToken :: Lens' GetResourceCollectionResponse (Maybe Text)

The pagination token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

getResourceCollectionResponse_httpStatus :: Lens' GetResourceCollectionResponse Int

The response's http status code.

ListEvents

listEvents_nextToken :: Lens' ListEvents (Maybe Text)

The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

listEvents_maxResults :: Lens' ListEvents (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

listEvents_filters :: Lens' ListEvents ListEventsFilters

A ListEventsFilters object used to specify which events to return.

listEventsResponse_nextToken :: Lens' ListEventsResponse (Maybe Text)

The pagination token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

listEventsResponse_httpStatus :: Lens' ListEventsResponse Int

The response's http status code.

listEventsResponse_events :: Lens' ListEventsResponse [Event]

A list of the requested events.

UpdateResourceCollection

updateResourceCollection_action :: Lens' UpdateResourceCollection UpdateResourceCollectionAction

Specifies if the resource collection in the request is added or deleted to the resource collection.

updateResourceCollection_resourceCollection :: Lens' UpdateResourceCollection UpdateResourceCollectionFilter

Undocumented member.

updateResourceCollectionResponse_httpStatus :: Lens' UpdateResourceCollectionResponse Int

The response's http status code.

StartCostEstimation

startCostEstimation_clientToken :: Lens' StartCostEstimation (Maybe Text)

The idempotency token used to identify each cost estimate request.

startCostEstimation_resourceCollection :: Lens' StartCostEstimation CostEstimationResourceCollectionFilter

The collection of AWS resources used to create a monthly DevOps Guru cost estimate.

startCostEstimationResponse_httpStatus :: Lens' StartCostEstimationResponse Int

The response's http status code.

ListRecommendations

listRecommendations_locale :: Lens' ListRecommendations (Maybe Locale)

A locale that specifies the language to use for recommendations.

listRecommendations_nextToken :: Lens' ListRecommendations (Maybe Text)

The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

listRecommendations_insightId :: Lens' ListRecommendations Text

The ID of the requested insight.

listRecommendationsResponse_nextToken :: Lens' ListRecommendationsResponse (Maybe Text)

The pagination token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

listRecommendationsResponse_recommendations :: Lens' ListRecommendationsResponse (Maybe [Recommendation])

An array of the requested recommendations.

listRecommendationsResponse_httpStatus :: Lens' ListRecommendationsResponse Int

The response's http status code.

DescribeAccountHealth

describeAccountHealthResponse_httpStatus :: Lens' DescribeAccountHealthResponse Int

The response's http status code.

describeAccountHealthResponse_openReactiveInsights :: Lens' DescribeAccountHealthResponse Int

An integer that specifies the number of open reactive insights in your AWS account.

describeAccountHealthResponse_openProactiveInsights :: Lens' DescribeAccountHealthResponse Int

An integer that specifies the number of open proactive insights in your AWS account.

describeAccountHealthResponse_metricsAnalyzed :: Lens' DescribeAccountHealthResponse Int

An integer that specifies the number of metrics that have been analyzed in your AWS account.

describeAccountHealthResponse_resourceHours :: Lens' DescribeAccountHealthResponse Integer

The number of Amazon DevOps Guru resource analysis hours billed to the current AWS account in the last hour.

DescribeInsight

describeInsight_id :: Lens' DescribeInsight Text

The ID of the insight.

describeInsightResponse_proactiveInsight :: Lens' DescribeInsightResponse (Maybe ProactiveInsight)

A ProactiveInsight object that represents the requested insight.

describeInsightResponse_reactiveInsight :: Lens' DescribeInsightResponse (Maybe ReactiveInsight)

A ReactiveInsight object that represents the requested insight.

describeInsightResponse_httpStatus :: Lens' DescribeInsightResponse Int

The response's http status code.

GetCostEstimation

getCostEstimation_nextToken :: Lens' GetCostEstimation (Maybe Text)

The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

getCostEstimationResponse_status :: Lens' GetCostEstimationResponse (Maybe CostEstimationStatus)

The status of creating this cost estimate. If it's still in progress, the status ONGOING is returned. If it is finished, the status COMPLETED is returned.

getCostEstimationResponse_resourceCollection :: Lens' GetCostEstimationResponse (Maybe CostEstimationResourceCollectionFilter)

The collection of the AWS resources used to create your monthly DevOps Guru cost estimate.

getCostEstimationResponse_timeRange :: Lens' GetCostEstimationResponse (Maybe CostEstimationTimeRange)

The start and end time of the cost estimation.

getCostEstimationResponse_costs :: Lens' GetCostEstimationResponse (Maybe [ServiceResourceCost])

An array of ResourceCost objects that each contains details about the monthly cost estimate to analyze one of your AWS resources.

getCostEstimationResponse_nextToken :: Lens' GetCostEstimationResponse (Maybe Text)

The pagination token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

getCostEstimationResponse_totalCost :: Lens' GetCostEstimationResponse (Maybe Double)

The estimated monthly cost to analyze the AWS resources. This value is the sum of the estimated costs to analyze each resource in the Costs object in this response.

getCostEstimationResponse_httpStatus :: Lens' GetCostEstimationResponse Int

The response's http status code.

Types

AnomalyReportedTimeRange

anomalyReportedTimeRange_closeTime :: Lens' AnomalyReportedTimeRange (Maybe UTCTime)

The time when an anomaly is closed.

anomalyReportedTimeRange_openTime :: Lens' AnomalyReportedTimeRange UTCTime

The time when an anomaly is opened.

AnomalySourceDetails

anomalySourceDetails_cloudWatchMetrics :: Lens' AnomalySourceDetails (Maybe [CloudWatchMetricsDetail])

An array of CloudWatchMetricsDetail object that contains information about the analyzed metrics that displayed anomalous behavior.

AnomalyTimeRange

anomalyTimeRange_endTime :: Lens' AnomalyTimeRange (Maybe UTCTime)

The time when the anomalous behavior ended.

anomalyTimeRange_startTime :: Lens' AnomalyTimeRange UTCTime

The time when the anomalous behavior started.

CloudFormationCollection

cloudFormationCollection_stackNames :: Lens' CloudFormationCollection (Maybe [Text])

An array of CloudFormation stack names.

CloudFormationCollectionFilter

cloudFormationCollectionFilter_stackNames :: Lens' CloudFormationCollectionFilter (Maybe [Text])

An array of CloudFormation stack names.

CloudFormationCostEstimationResourceCollectionFilter

cloudFormationCostEstimationResourceCollectionFilter_stackNames :: Lens' CloudFormationCostEstimationResourceCollectionFilter (Maybe (NonEmpty Text))

An array of CloudFormation stack names. Its size is fixed at 1 item.

CloudFormationHealth

cloudFormationHealth_insight :: Lens' CloudFormationHealth (Maybe InsightHealth)

Information about the health of the AWS resources in your account that are specified by an AWS CloudFormation stack, including the number of open proactive, open reactive insights, and the Mean Time to Recover (MTTR) of closed insights.

cloudFormationHealth_stackName :: Lens' CloudFormationHealth (Maybe Text)

The name of the CloudFormation stack.

CloudWatchMetricsDetail

cloudWatchMetricsDetail_period :: Lens' CloudWatchMetricsDetail (Maybe Int)

The length of time associated with the CloudWatch metric in number of seconds.

cloudWatchMetricsDetail_metricName :: Lens' CloudWatchMetricsDetail (Maybe Text)

The name of the CloudWatch metric.

cloudWatchMetricsDetail_namespace :: Lens' CloudWatchMetricsDetail (Maybe Text)

The namespace of the CloudWatch metric. A namespace is a container for CloudWatch metrics.

cloudWatchMetricsDetail_stat :: Lens' CloudWatchMetricsDetail (Maybe CloudWatchMetricsStat)

The type of statistic associated with the CloudWatch metric. For more information, see Statistics in the Amazon CloudWatch User Guide.

cloudWatchMetricsDetail_dimensions :: Lens' CloudWatchMetricsDetail (Maybe [CloudWatchMetricsDimension])

An array of CloudWatch dimensions associated with

cloudWatchMetricsDetail_unit :: Lens' CloudWatchMetricsDetail (Maybe Text)

The unit of measure used for the CloudWatch metric. For example, Bytes, Seconds, Count, and Percent.

CloudWatchMetricsDimension

cloudWatchMetricsDimension_value :: Lens' CloudWatchMetricsDimension (Maybe Text)

The value of the CloudWatch dimension.

cloudWatchMetricsDimension_name :: Lens' CloudWatchMetricsDimension (Maybe Text)

The name of the CloudWatch dimension.

CostEstimationResourceCollectionFilter

costEstimationResourceCollectionFilter_cloudFormation :: Lens' CostEstimationResourceCollectionFilter (Maybe CloudFormationCostEstimationResourceCollectionFilter)

An object that specifies the CloudFormation stack that defines the AWS resources used to create a monthly estimate for DevOps Guru.

CostEstimationTimeRange

costEstimationTimeRange_startTime :: Lens' CostEstimationTimeRange (Maybe UTCTime)

The start time of the cost estimation.

costEstimationTimeRange_endTime :: Lens' CostEstimationTimeRange (Maybe UTCTime)

The end time of the cost estimation.

EndTimeRange

endTimeRange_fromTime :: Lens' EndTimeRange (Maybe UTCTime)

The earliest end time in the time range.

endTimeRange_toTime :: Lens' EndTimeRange (Maybe UTCTime)

The latest end time in the time range.

Event

event_resourceCollection :: Lens' Event (Maybe ResourceCollection)

Undocumented member.

event_eventClass :: Lens' Event (Maybe EventClass)

The class of the event. The class specifies what the event is related to, such as an infrastructure change, a deployment, or a schema change.

event_time :: Lens' Event (Maybe UTCTime)

A Timestamp that specifies the time the event occurred.

event_resources :: Lens' Event (Maybe [EventResource])

An EventResource object that contains information about the resource that emitted the event.

event_name :: Lens' Event (Maybe Text)

The name of the event.

event_id :: Lens' Event (Maybe Text)

The ID of the event.

event_dataSource :: Lens' Event (Maybe EventDataSource)

The source, AWS_CLOUD_TRAIL or AWS_CODE_DEPLOY, where DevOps Guru analysis found the event.

event_eventSource :: Lens' Event (Maybe Text)

The AWS source that emitted the event.

EventResource

eventResource_arn :: Lens' EventResource (Maybe Text)

The Amazon Resource Name (ARN) of the resource that emitted an event.

eventResource_name :: Lens' EventResource (Maybe Text)

The name of the resource that emitted an event.

eventResource_type :: Lens' EventResource (Maybe Text)

The type of resource that emitted an event.

EventTimeRange

eventTimeRange_fromTime :: Lens' EventTimeRange UTCTime

The time when the event started.

eventTimeRange_toTime :: Lens' EventTimeRange UTCTime

The time when the event ended.

InsightFeedback

insightFeedback_id :: Lens' InsightFeedback (Maybe Text)

The insight feedback ID.

insightFeedback_feedback :: Lens' InsightFeedback (Maybe InsightFeedbackOption)

The feedback provided by the customer.

InsightHealth

insightHealth_meanTimeToRecoverInMilliseconds :: Lens' InsightHealth (Maybe Integer)

The Meant Time to Recover (MTTR) for the insight.

insightHealth_openReactiveInsights :: Lens' InsightHealth (Maybe Int)

The number of open reactive insights.

insightHealth_openProactiveInsights :: Lens' InsightHealth (Maybe Int)

The number of open proactive insights.

InsightTimeRange

insightTimeRange_endTime :: Lens' InsightTimeRange (Maybe UTCTime)

The time when the behavior described in an insight ended.

insightTimeRange_startTime :: Lens' InsightTimeRange UTCTime

The time when the behavior described in an insight started.

ListEventsFilters

listEventsFilters_resourceCollection :: Lens' ListEventsFilters (Maybe ResourceCollection)

Undocumented member.

listEventsFilters_eventClass :: Lens' ListEventsFilters (Maybe EventClass)

The class of the events you want to filter for, such as an infrastructure change, a deployment, or a schema change.

listEventsFilters_insightId :: Lens' ListEventsFilters (Maybe Text)

An ID of an insight that is related to the events you want to filter for.

listEventsFilters_dataSource :: Lens' ListEventsFilters (Maybe EventDataSource)

The source, AWS_CLOUD_TRAIL or AWS_CODE_DEPLOY, of the events you want returned.

listEventsFilters_eventTimeRange :: Lens' ListEventsFilters (Maybe EventTimeRange)

A time range during which you want the filtered events to have occurred.

listEventsFilters_eventSource :: Lens' ListEventsFilters (Maybe Text)

The AWS source that emitted the events you want to filter for.

ListInsightsAnyStatusFilter

listInsightsAnyStatusFilter_type :: Lens' ListInsightsAnyStatusFilter InsightType

Use to filter for either REACTIVE or PROACTIVE insights.

listInsightsAnyStatusFilter_startTimeRange :: Lens' ListInsightsAnyStatusFilter StartTimeRange

A time range used to specify when the behavior of the filtered insights started.

ListInsightsClosedStatusFilter

listInsightsClosedStatusFilter_type :: Lens' ListInsightsClosedStatusFilter InsightType

Use to filter for either REACTIVE or PROACTIVE insights.

listInsightsClosedStatusFilter_endTimeRange :: Lens' ListInsightsClosedStatusFilter EndTimeRange

A time range used to specify when the behavior of the filtered insights ended.

ListInsightsOngoingStatusFilter

listInsightsOngoingStatusFilter_type :: Lens' ListInsightsOngoingStatusFilter InsightType

Use to filter for either REACTIVE or PROACTIVE insights.

ListInsightsStatusFilter

listInsightsStatusFilter_closed :: Lens' ListInsightsStatusFilter (Maybe ListInsightsClosedStatusFilter)

A ListInsightsClosedStatusFilter that specifies closed insights that are either REACTIVE or PROACTIVE.

listInsightsStatusFilter_ongoing :: Lens' ListInsightsStatusFilter (Maybe ListInsightsOngoingStatusFilter)

A ListInsightsAnyStatusFilter that specifies ongoing insights that are either REACTIVE or PROACTIVE.

listInsightsStatusFilter_any :: Lens' ListInsightsStatusFilter (Maybe ListInsightsAnyStatusFilter)

A ListInsightsAnyStatusFilter that specifies insights of any status that are either REACTIVE or PROACTIVE.

NotificationChannel

notificationChannel_config :: Lens' NotificationChannel (Maybe NotificationChannelConfig)

A NotificationChannelConfig object that contains information about configured notification channels.

notificationChannel_id :: Lens' NotificationChannel (Maybe Text)

The ID of a notification channel.

NotificationChannelConfig

notificationChannelConfig_sns :: Lens' NotificationChannelConfig SnsChannelConfig

Information about a notification channel configured in DevOps Guru to send notifications when insights are created.

If you use an Amazon SNS topic in another account, you must attach a policy to it that grants DevOps Guru permission to it notifications. DevOps Guru adds the required policy on your behalf to send notifications using Amazon SNS in your account. For more information, see Permissions for cross account Amazon SNS topics.

If you use an Amazon SNS topic that is encrypted by an AWS Key Management Service customer-managed key (CMK), then you must add permissions to the CMK. For more information, see Permissions for AWS KMS–encrypted Amazon SNS topics.

OpsCenterIntegration

opsCenterIntegration_optInStatus :: Lens' OpsCenterIntegration (Maybe OptInStatus)

Specifies if DevOps Guru is enabled to create an AWS Systems Manager OpsItem for each created insight.

OpsCenterIntegrationConfig

opsCenterIntegrationConfig_optInStatus :: Lens' OpsCenterIntegrationConfig (Maybe OptInStatus)

Specifies if DevOps Guru is enabled to create an AWS Systems Manager OpsItem for each created insight.

PredictionTimeRange

predictionTimeRange_endTime :: Lens' PredictionTimeRange (Maybe UTCTime)

The time when the behavior in a proactive insight is expected to end.

predictionTimeRange_startTime :: Lens' PredictionTimeRange UTCTime

The time range during which a metric limit is expected to be exceeded. This applies to proactive insights only.

ProactiveAnomaly

proactiveAnomaly_anomalyReportedTimeRange :: Lens' ProactiveAnomaly (Maybe AnomalyReportedTimeRange)

A AnomalyReportedTimeRange object that specifies the time range between when the anomaly is opened and the time when it is closed.

proactiveAnomaly_status :: Lens' ProactiveAnomaly (Maybe AnomalyStatus)

The status of a proactive anomaly.

proactiveAnomaly_resourceCollection :: Lens' ProactiveAnomaly (Maybe ResourceCollection)

Undocumented member.

proactiveAnomaly_severity :: Lens' ProactiveAnomaly (Maybe AnomalySeverity)

The severity of a proactive anomaly.

proactiveAnomaly_updateTime :: Lens' ProactiveAnomaly (Maybe UTCTime)

The time of the anomaly's most recent update.

proactiveAnomaly_sourceDetails :: Lens' ProactiveAnomaly (Maybe AnomalySourceDetails)

Details about the source of the analyzed operational data that triggered the anomaly. The one supported source is Amazon CloudWatch metrics.

proactiveAnomaly_predictionTimeRange :: Lens' ProactiveAnomaly (Maybe PredictionTimeRange)

Undocumented member.

proactiveAnomaly_limit :: Lens' ProactiveAnomaly (Maybe Double)

A threshold that was exceeded by behavior in analyzed resources. Exceeding this threshold is related to the anomalous behavior that generated this anomaly.

proactiveAnomaly_id :: Lens' ProactiveAnomaly (Maybe Text)

The ID of a proactive anomaly.

proactiveAnomaly_associatedInsightId :: Lens' ProactiveAnomaly (Maybe Text)

The ID of the insight that contains this anomaly. An insight is composed of related anomalies.

proactiveAnomaly_anomalyTimeRange :: Lens' ProactiveAnomaly (Maybe AnomalyTimeRange)

Undocumented member.

ProactiveAnomalySummary

proactiveAnomalySummary_anomalyReportedTimeRange :: Lens' ProactiveAnomalySummary (Maybe AnomalyReportedTimeRange)

A AnomalyReportedTimeRange object that specifies the time range between when the anomaly is opened and the time when it is closed.

proactiveAnomalySummary_status :: Lens' ProactiveAnomalySummary (Maybe AnomalyStatus)

The status of the anomaly.

proactiveAnomalySummary_resourceCollection :: Lens' ProactiveAnomalySummary (Maybe ResourceCollection)

Undocumented member.

proactiveAnomalySummary_severity :: Lens' ProactiveAnomalySummary (Maybe AnomalySeverity)

The severity of the anomaly.

proactiveAnomalySummary_updateTime :: Lens' ProactiveAnomalySummary (Maybe UTCTime)

The time of the anomaly's most recent update.

proactiveAnomalySummary_sourceDetails :: Lens' ProactiveAnomalySummary (Maybe AnomalySourceDetails)

Details about the source of the analyzed operational data that triggered the anomaly. The one supported source is Amazon CloudWatch metrics.

proactiveAnomalySummary_predictionTimeRange :: Lens' ProactiveAnomalySummary (Maybe PredictionTimeRange)

Undocumented member.

proactiveAnomalySummary_limit :: Lens' ProactiveAnomalySummary (Maybe Double)

A threshold that was exceeded by behavior in analyzed resources. Exceeding this threshold is related to the anomalous behavior that generated this anomaly.

proactiveAnomalySummary_id :: Lens' ProactiveAnomalySummary (Maybe Text)

The ID of the anomaly.

proactiveAnomalySummary_associatedInsightId :: Lens' ProactiveAnomalySummary (Maybe Text)

The ID of the insight that contains this anomaly. An insight is composed of related anomalies.

proactiveAnomalySummary_anomalyTimeRange :: Lens' ProactiveAnomalySummary (Maybe AnomalyTimeRange)

Undocumented member.

ProactiveInsight

proactiveInsight_status :: Lens' ProactiveInsight (Maybe InsightStatus)

The status of the proactive insight.

proactiveInsight_resourceCollection :: Lens' ProactiveInsight (Maybe ResourceCollection)

Undocumented member.

proactiveInsight_severity :: Lens' ProactiveInsight (Maybe InsightSeverity)

The severity of the proactive insight.

proactiveInsight_ssmOpsItemId :: Lens' ProactiveInsight (Maybe Text)

The ID of the AWS System Manager OpsItem created for this insight. You must enable the creation of OpstItems insights before they are created for each insight.

proactiveInsight_insightTimeRange :: Lens' ProactiveInsight (Maybe InsightTimeRange)

Undocumented member.

proactiveInsight_name :: Lens' ProactiveInsight (Maybe Text)

The name of the proactive insight.

proactiveInsight_predictionTimeRange :: Lens' ProactiveInsight (Maybe PredictionTimeRange)

Undocumented member.

proactiveInsight_id :: Lens' ProactiveInsight (Maybe Text)

The ID of the proactive insight.

ProactiveInsightSummary

proactiveInsightSummary_status :: Lens' ProactiveInsightSummary (Maybe InsightStatus)

The status of the proactive insight.

proactiveInsightSummary_resourceCollection :: Lens' ProactiveInsightSummary (Maybe ResourceCollection)

Undocumented member.

proactiveInsightSummary_severity :: Lens' ProactiveInsightSummary (Maybe InsightSeverity)

The severity of the proactive insight.

proactiveInsightSummary_insightTimeRange :: Lens' ProactiveInsightSummary (Maybe InsightTimeRange)

Undocumented member.

proactiveInsightSummary_name :: Lens' ProactiveInsightSummary (Maybe Text)

The name of the proactive insight.

proactiveInsightSummary_predictionTimeRange :: Lens' ProactiveInsightSummary (Maybe PredictionTimeRange)

Undocumented member.

proactiveInsightSummary_id :: Lens' ProactiveInsightSummary (Maybe Text)

The ID of the proactive insight.

proactiveInsightSummary_serviceCollection :: Lens' ProactiveInsightSummary (Maybe ServiceCollection)

A collection of the names of AWS services.

ReactiveAnomaly

reactiveAnomaly_anomalyReportedTimeRange :: Lens' ReactiveAnomaly (Maybe AnomalyReportedTimeRange)

A AnomalyReportedTimeRange object that specifies the time range between when the anomaly is opened and the time when it is closed.

reactiveAnomaly_status :: Lens' ReactiveAnomaly (Maybe AnomalyStatus)

The status of the anomaly.

reactiveAnomaly_resourceCollection :: Lens' ReactiveAnomaly (Maybe ResourceCollection)

Undocumented member.

reactiveAnomaly_severity :: Lens' ReactiveAnomaly (Maybe AnomalySeverity)

The severity of the anomaly.

reactiveAnomaly_sourceDetails :: Lens' ReactiveAnomaly (Maybe AnomalySourceDetails)

Details about the source of the analyzed operational data that triggered the anomaly. The one supported source is Amazon CloudWatch metrics.

reactiveAnomaly_id :: Lens' ReactiveAnomaly (Maybe Text)

The ID of the reactive anomaly.

reactiveAnomaly_associatedInsightId :: Lens' ReactiveAnomaly (Maybe Text)

The ID of the insight that contains this anomaly. An insight is composed of related anomalies.

reactiveAnomaly_anomalyTimeRange :: Lens' ReactiveAnomaly (Maybe AnomalyTimeRange)

Undocumented member.

ReactiveAnomalySummary

reactiveAnomalySummary_anomalyReportedTimeRange :: Lens' ReactiveAnomalySummary (Maybe AnomalyReportedTimeRange)

A AnomalyReportedTimeRange object that specifies the time range between when the anomaly is opened and the time when it is closed.

reactiveAnomalySummary_status :: Lens' ReactiveAnomalySummary (Maybe AnomalyStatus)

The status of the reactive anomaly.

reactiveAnomalySummary_resourceCollection :: Lens' ReactiveAnomalySummary (Maybe ResourceCollection)

Undocumented member.

reactiveAnomalySummary_severity :: Lens' ReactiveAnomalySummary (Maybe AnomalySeverity)

The severity of the reactive anomaly.

reactiveAnomalySummary_sourceDetails :: Lens' ReactiveAnomalySummary (Maybe AnomalySourceDetails)

Details about the source of the analyzed operational data that triggered the anomaly. The one supported source is Amazon CloudWatch metrics.

reactiveAnomalySummary_id :: Lens' ReactiveAnomalySummary (Maybe Text)

The ID of the reactive anomaly.

reactiveAnomalySummary_associatedInsightId :: Lens' ReactiveAnomalySummary (Maybe Text)

The ID of the insight that contains this anomaly. An insight is composed of related anomalies.

reactiveAnomalySummary_anomalyTimeRange :: Lens' ReactiveAnomalySummary (Maybe AnomalyTimeRange)

Undocumented member.

ReactiveInsight

reactiveInsight_status :: Lens' ReactiveInsight (Maybe InsightStatus)

The status of a reactive insight.

reactiveInsight_resourceCollection :: Lens' ReactiveInsight (Maybe ResourceCollection)

Undocumented member.

reactiveInsight_severity :: Lens' ReactiveInsight (Maybe InsightSeverity)

The severity of a reactive insight.

reactiveInsight_ssmOpsItemId :: Lens' ReactiveInsight (Maybe Text)

The ID of the AWS System Manager OpsItem created for this insight. You must enable the creation of OpstItems insights before they are created for each insight.

reactiveInsight_insightTimeRange :: Lens' ReactiveInsight (Maybe InsightTimeRange)

Undocumented member.

reactiveInsight_name :: Lens' ReactiveInsight (Maybe Text)

The name of a reactive insight.

reactiveInsight_id :: Lens' ReactiveInsight (Maybe Text)

The ID of a reactive insight.

ReactiveInsightSummary

reactiveInsightSummary_status :: Lens' ReactiveInsightSummary (Maybe InsightStatus)

The status of a reactive insight.

reactiveInsightSummary_resourceCollection :: Lens' ReactiveInsightSummary (Maybe ResourceCollection)

Undocumented member.

reactiveInsightSummary_severity :: Lens' ReactiveInsightSummary (Maybe InsightSeverity)

The severity of a reactive insight.

reactiveInsightSummary_insightTimeRange :: Lens' ReactiveInsightSummary (Maybe InsightTimeRange)

Undocumented member.

reactiveInsightSummary_name :: Lens' ReactiveInsightSummary (Maybe Text)

The name of a reactive insight.

reactiveInsightSummary_id :: Lens' ReactiveInsightSummary (Maybe Text)

The ID of a reactive summary.

reactiveInsightSummary_serviceCollection :: Lens' ReactiveInsightSummary (Maybe ServiceCollection)

A collection of the names of AWS services.

Recommendation

recommendation_link :: Lens' Recommendation (Maybe Text)

A hyperlink to information to help you address the problem.

recommendation_relatedAnomalies :: Lens' Recommendation (Maybe [RecommendationRelatedAnomaly])

Anomalies that are related to the problem. Use these Anomalies to learn more about what's happening and to help address the issue.

recommendation_reason :: Lens' Recommendation (Maybe Text)

The reason DevOps Guru flagged the anomalous behavior as a problem.

recommendation_name :: Lens' Recommendation (Maybe Text)

The name of the recommendation.

recommendation_relatedEvents :: Lens' Recommendation (Maybe [RecommendationRelatedEvent])

Events that are related to the problem. Use these events to learn more about what's happening and to help address the issue.

recommendation_description :: Lens' Recommendation (Maybe Text)

A description of the problem.

RecommendationRelatedAnomaly

recommendationRelatedAnomaly_resources :: Lens' RecommendationRelatedAnomaly (Maybe [RecommendationRelatedAnomalyResource])

An array of objects that represent resources in which DevOps Guru detected anomalous behavior. Each object contains the name and type of the resource.

recommendationRelatedAnomaly_sourceDetails :: Lens' RecommendationRelatedAnomaly (Maybe [RecommendationRelatedAnomalySourceDetail])

Information about where the anomalous behavior related the recommendation was found. For example, details in Amazon CloudWatch metrics.

RecommendationRelatedAnomalyResource

recommendationRelatedAnomalyResource_name :: Lens' RecommendationRelatedAnomalyResource (Maybe Text)

The name of the resource.

recommendationRelatedAnomalyResource_type :: Lens' RecommendationRelatedAnomalyResource (Maybe Text)

The type of the resource.

RecommendationRelatedAnomalySourceDetail

recommendationRelatedAnomalySourceDetail_cloudWatchMetrics :: Lens' RecommendationRelatedAnomalySourceDetail (Maybe [RecommendationRelatedCloudWatchMetricsSourceDetail])

An array of CloudWatchMetricsDetail objects that contains information about the analyzed metrics that displayed anomalous behavior.

RecommendationRelatedCloudWatchMetricsSourceDetail

recommendationRelatedCloudWatchMetricsSourceDetail_metricName :: Lens' RecommendationRelatedCloudWatchMetricsSourceDetail (Maybe Text)

The name of the CloudWatch metric.

recommendationRelatedCloudWatchMetricsSourceDetail_namespace :: Lens' RecommendationRelatedCloudWatchMetricsSourceDetail (Maybe Text)

The namespace of the CloudWatch metric. A namespace is a container for CloudWatch metrics.

RecommendationRelatedEvent

recommendationRelatedEvent_resources :: Lens' RecommendationRelatedEvent (Maybe [RecommendationRelatedEventResource])

A ResourceCollection object that contains arrays of the names of AWS CloudFormation stacks. You can specify up to 500 AWS CloudFormation stacks.

recommendationRelatedEvent_name :: Lens' RecommendationRelatedEvent (Maybe Text)

The name of the event. This corresponds to the Name field in an Event object.

RecommendationRelatedEventResource

recommendationRelatedEventResource_name :: Lens' RecommendationRelatedEventResource (Maybe Text)

The name of the resource that emitted the event. This corresponds to the Name field in an EventResource object.

recommendationRelatedEventResource_type :: Lens' RecommendationRelatedEventResource (Maybe Text)

The type of the resource that emitted the event. This corresponds to the Type field in an EventResource object.

ResourceCollection

resourceCollection_cloudFormation :: Lens' ResourceCollection (Maybe CloudFormationCollection)

An array of the names of AWS CloudFormation stacks. The stacks define AWS resources that DevOps Guru analyzes. You can specify up to 500 AWS CloudFormation stacks.

ResourceCollectionFilter

resourceCollectionFilter_cloudFormation :: Lens' ResourceCollectionFilter (Maybe CloudFormationCollectionFilter)

Information about AWS CloudFormation stacks. You can use up to 500 stacks to specify which AWS resources in your account to analyze. For more information, see Stacks in the AWS CloudFormation User Guide.

SearchInsightsFilters

searchInsightsFilters_resourceCollection :: Lens' SearchInsightsFilters (Maybe ResourceCollection)

Undocumented member.

searchInsightsFilters_statuses :: Lens' SearchInsightsFilters (Maybe [InsightStatus])

An array of status values used to search for insights.

searchInsightsFilters_severities :: Lens' SearchInsightsFilters (Maybe [InsightSeverity])

An array of severity values used to search for insights.

searchInsightsFilters_serviceCollection :: Lens' SearchInsightsFilters (Maybe ServiceCollection)

A collection of the names of AWS services.

ServiceCollection

serviceCollection_serviceNames :: Lens' ServiceCollection (Maybe [ServiceName])

An array of strings that each specifies the name of an AWS service.

ServiceHealth

serviceHealth_insight :: Lens' ServiceHealth (Maybe ServiceInsightHealth)

Represents the health of an AWS service. This is a ServiceInsightHealth that contains the number of open proactive and reactive insights for this service.

serviceHealth_serviceName :: Lens' ServiceHealth (Maybe ServiceName)

The name of the AWS service.

ServiceInsightHealth

serviceInsightHealth_openReactiveInsights :: Lens' ServiceInsightHealth (Maybe Int)

The number of open reactive insights in the AWS service

serviceInsightHealth_openProactiveInsights :: Lens' ServiceInsightHealth (Maybe Int)

The number of open proactive insights in the AWS service

ServiceIntegrationConfig

serviceIntegrationConfig_opsCenter :: Lens' ServiceIntegrationConfig (Maybe OpsCenterIntegration)

Information about whether DevOps Guru is configured to create an OpsItem in AWS Systems Manager OpsCenter for each created insight.

ServiceResourceCost

serviceResourceCost_state :: Lens' ServiceResourceCost (Maybe CostEstimationServiceResourceState)

The state of the resource. The resource is ACTIVE if it produces metrics, events, or logs within an hour, otherwise it is INACTIVE. You pay for the number of active AWS resource hours analyzed for each resource. Inactive resources are not charged.

serviceResourceCost_unitCost :: Lens' ServiceResourceCost (Maybe Double)

The price per hour to analyze the resources in the service. For more information, see Estimate your Amazon DevOps Guru costs and Amazon DevOps Guru pricing.

serviceResourceCost_count :: Lens' ServiceResourceCost (Maybe Int)

The number of active resources analyzed for this service to create a monthly cost estimate.

serviceResourceCost_cost :: Lens' ServiceResourceCost (Maybe Double)

The total estimated monthly cost to analyze the active resources for this resource.

serviceResourceCost_type :: Lens' ServiceResourceCost (Maybe Text)

The type of the AWS resource.

SnsChannelConfig

snsChannelConfig_topicArn :: Lens' SnsChannelConfig (Maybe Text)

The Amazon Resource Name (ARN) of an Amazon Simple Notification Service topic.

StartTimeRange

startTimeRange_fromTime :: Lens' StartTimeRange (Maybe UTCTime)

The start time of the time range.

startTimeRange_toTime :: Lens' StartTimeRange (Maybe UTCTime)

The end time of the time range.

UpdateCloudFormationCollectionFilter

updateCloudFormationCollectionFilter_stackNames :: Lens' UpdateCloudFormationCollectionFilter (Maybe [Text])

An array of the names of the AWS CloudFormation stacks to update. You can specify up to 500 AWS CloudFormation stacks.

UpdateResourceCollectionFilter

updateResourceCollectionFilter_cloudFormation :: Lens' UpdateResourceCollectionFilter (Maybe UpdateCloudFormationCollectionFilter)

An collection of AWS CloudFormation stacks. You can specify up to 500 AWS CloudFormation stacks.

UpdateServiceIntegrationConfig

updateServiceIntegrationConfig_opsCenter :: Lens' UpdateServiceIntegrationConfig (Maybe OpsCenterIntegrationConfig)

Undocumented member.