Network.AWS.DevOpsGuru.ListEvents

Description

Returns a list of the events emitted by the resources that are evaluated by DevOps Guru. You can use filters to specify which events are returned.

This operation returns paginated results.

Creating a Request

data ListEvents

See: newListEvents smart constructor.

Constructors

ListEvents'

Fields

nextToken :: Maybe Text

The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

maxResults :: Maybe Natural

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

filters :: ListEventsFilters

A ListEventsFilters object used to specify which events to return.

newListEvents

Arguments

:: ListEventsFilters
$sel:filters:ListEvents'

-> ListEvents

Create a value of ListEvents with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListEvents', listEvents_nextToken - The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

$sel:maxResults:ListEvents', listEvents_maxResults - The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

$sel:filters:ListEvents', listEvents_filters - A ListEventsFilters object used to specify which events to return.

Request Lenses

listEvents_nextToken :: Lens' ListEvents (Maybe Text)

The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

listEvents_maxResults :: Lens' ListEvents (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

listEvents_filters :: Lens' ListEvents ListEventsFilters

A ListEventsFilters object used to specify which events to return.

Destructuring the Response

data ListEventsResponse

See: newListEventsResponse smart constructor.

Constructors

ListEventsResponse'

Fields

nextToken :: Maybe Text

The pagination token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

httpStatus :: Int

The response's http status code.

events :: [Event]

A list of the requested events.

newListEventsResponse

Arguments

:: Int
$sel:httpStatus:ListEventsResponse'

-> ListEventsResponse

Create a value of ListEventsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListEvents', listEventsResponse_nextToken - The pagination token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

$sel:httpStatus:ListEventsResponse', listEventsResponse_httpStatus - The response's http status code.

$sel:events:ListEventsResponse', listEventsResponse_events - A list of the requested events.

Response Lenses

listEventsResponse_nextToken :: Lens' ListEventsResponse (Maybe Text)

The pagination token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

listEventsResponse_httpStatus :: Lens' ListEventsResponse Int

The response's http status code.

listEventsResponse_events :: Lens' ListEventsResponse [Event]

A list of the requested events.