Network.AWS.DevOpsGuru.Types.AnomalyTimeRange

Description

Documentation

data AnomalyTimeRange

A time range that specifies when the observed unusual behavior in an anomaly started and ended. This is different from AnomalyReportedTimeRange, which specifies the time range when DevOps Guru opens and then closes an anomaly.

See: newAnomalyTimeRange smart constructor.

Constructors

AnomalyTimeRange'

Fields

endTime :: Maybe POSIX

The time when the anomalous behavior ended.

startTime :: POSIX

The time when the anomalous behavior started.

newAnomalyTimeRange

Arguments

:: UTCTime
$sel:startTime:AnomalyTimeRange'

-> AnomalyTimeRange

Create a value of AnomalyTimeRange with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:endTime:AnomalyTimeRange', anomalyTimeRange_endTime - The time when the anomalous behavior ended.

$sel:startTime:AnomalyTimeRange', anomalyTimeRange_startTime - The time when the anomalous behavior started.

anomalyTimeRange_endTime :: Lens' AnomalyTimeRange (Maybe UTCTime)

The time when the anomalous behavior ended.

anomalyTimeRange_startTime :: Lens' AnomalyTimeRange UTCTime

The time when the anomalous behavior started.