Network.AWS.DevOpsGuru.Types.InsightHealth

Description

Documentation

data InsightHealth

Information about the number of open reactive and proactive insights that can be used to gauge the health of your system.

See: newInsightHealth smart constructor.

Constructors

InsightHealth'

Fields

meanTimeToRecoverInMilliseconds :: Maybe Integer

The Meant Time to Recover (MTTR) for the insight.

openReactiveInsights :: Maybe Int

The number of open reactive insights.

openProactiveInsights :: Maybe Int

The number of open proactive insights.

newInsightHealth :: InsightHealth

Create a value of InsightHealth with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:meanTimeToRecoverInMilliseconds:InsightHealth', insightHealth_meanTimeToRecoverInMilliseconds - The Meant Time to Recover (MTTR) for the insight.

$sel:openReactiveInsights:InsightHealth', insightHealth_openReactiveInsights - The number of open reactive insights.

$sel:openProactiveInsights:InsightHealth', insightHealth_openProactiveInsights - The number of open proactive insights.

insightHealth_meanTimeToRecoverInMilliseconds :: Lens' InsightHealth (Maybe Integer)

The Meant Time to Recover (MTTR) for the insight.

insightHealth_openReactiveInsights :: Lens' InsightHealth (Maybe Int)

The number of open reactive insights.

insightHealth_openProactiveInsights :: Lens' InsightHealth (Maybe Int)

The number of open proactive insights.