Network.AWS.DevOpsGuru.Types.ListInsightsAnyStatusFilter

Description

Documentation

data ListInsightsAnyStatusFilter

Used to filter for insights that have any status.

See: newListInsightsAnyStatusFilter smart constructor.

Constructors

ListInsightsAnyStatusFilter'

Fields

type' :: InsightType

Use to filter for either REACTIVE or PROACTIVE insights.

startTimeRange :: StartTimeRange

A time range used to specify when the behavior of the filtered insights started.

newListInsightsAnyStatusFilter

Arguments

:: InsightType
$sel:type':ListInsightsAnyStatusFilter'

-> StartTimeRange
$sel:startTimeRange:ListInsightsAnyStatusFilter'

-> ListInsightsAnyStatusFilter

Create a value of ListInsightsAnyStatusFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:type':ListInsightsAnyStatusFilter', listInsightsAnyStatusFilter_type - Use to filter for either REACTIVE or PROACTIVE insights.

$sel:startTimeRange:ListInsightsAnyStatusFilter', listInsightsAnyStatusFilter_startTimeRange - A time range used to specify when the behavior of the filtered insights started.

listInsightsAnyStatusFilter_type :: Lens' ListInsightsAnyStatusFilter InsightType

Use to filter for either REACTIVE or PROACTIVE insights.

listInsightsAnyStatusFilter_startTimeRange :: Lens' ListInsightsAnyStatusFilter StartTimeRange

A time range used to specify when the behavior of the filtered insights started.