Network.AWS.DevOpsGuru.Types.ProactiveAnomaly

Description

Documentation

data ProactiveAnomaly

Information about an anomaly. This object is returned by ListAnomalies.

See: newProactiveAnomaly smart constructor.

Constructors

ProactiveAnomaly'

Fields

anomalyReportedTimeRange :: Maybe AnomalyReportedTimeRange

A AnomalyReportedTimeRange object that specifies the time range between when the anomaly is opened and the time when it is closed.

status :: Maybe AnomalyStatus

The status of a proactive anomaly.

resourceCollection :: Maybe ResourceCollection

severity :: Maybe AnomalySeverity

The severity of a proactive anomaly.

updateTime :: Maybe POSIX

The time of the anomaly's most recent update.

sourceDetails :: Maybe AnomalySourceDetails

Details about the source of the analyzed operational data that triggered the anomaly. The one supported source is Amazon CloudWatch metrics.

predictionTimeRange :: Maybe PredictionTimeRange

limit :: Maybe Double

A threshold that was exceeded by behavior in analyzed resources. Exceeding this threshold is related to the anomalous behavior that generated this anomaly.

id :: Maybe Text

The ID of a proactive anomaly.

associatedInsightId :: Maybe Text

The ID of the insight that contains this anomaly. An insight is composed of related anomalies.

anomalyTimeRange :: Maybe AnomalyTimeRange

newProactiveAnomaly :: ProactiveAnomaly

Create a value of ProactiveAnomaly with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:anomalyReportedTimeRange:ProactiveAnomaly', proactiveAnomaly_anomalyReportedTimeRange - A AnomalyReportedTimeRange object that specifies the time range between when the anomaly is opened and the time when it is closed.

$sel:status:ProactiveAnomaly', proactiveAnomaly_status - The status of a proactive anomaly.

$sel:resourceCollection:ProactiveAnomaly', proactiveAnomaly_resourceCollection - Undocumented member.

$sel:severity:ProactiveAnomaly', proactiveAnomaly_severity - The severity of a proactive anomaly.

$sel:updateTime:ProactiveAnomaly', proactiveAnomaly_updateTime - The time of the anomaly's most recent update.

$sel:sourceDetails:ProactiveAnomaly', proactiveAnomaly_sourceDetails - Details about the source of the analyzed operational data that triggered the anomaly. The one supported source is Amazon CloudWatch metrics.

$sel:predictionTimeRange:ProactiveAnomaly', proactiveAnomaly_predictionTimeRange - Undocumented member.

$sel:limit:ProactiveAnomaly', proactiveAnomaly_limit - A threshold that was exceeded by behavior in analyzed resources. Exceeding this threshold is related to the anomalous behavior that generated this anomaly.

$sel:id:ProactiveAnomaly', proactiveAnomaly_id - The ID of a proactive anomaly.

$sel:associatedInsightId:ProactiveAnomaly', proactiveAnomaly_associatedInsightId - The ID of the insight that contains this anomaly. An insight is composed of related anomalies.

$sel:anomalyTimeRange:ProactiveAnomaly', proactiveAnomaly_anomalyTimeRange - Undocumented member.

proactiveAnomaly_anomalyReportedTimeRange :: Lens' ProactiveAnomaly (Maybe AnomalyReportedTimeRange)

A AnomalyReportedTimeRange object that specifies the time range between when the anomaly is opened and the time when it is closed.

proactiveAnomaly_status :: Lens' ProactiveAnomaly (Maybe AnomalyStatus)

The status of a proactive anomaly.

proactiveAnomaly_resourceCollection :: Lens' ProactiveAnomaly (Maybe ResourceCollection)

Undocumented member.

proactiveAnomaly_severity :: Lens' ProactiveAnomaly (Maybe AnomalySeverity)

The severity of a proactive anomaly.

proactiveAnomaly_updateTime :: Lens' ProactiveAnomaly (Maybe UTCTime)

The time of the anomaly's most recent update.

proactiveAnomaly_sourceDetails :: Lens' ProactiveAnomaly (Maybe AnomalySourceDetails)

Details about the source of the analyzed operational data that triggered the anomaly. The one supported source is Amazon CloudWatch metrics.

proactiveAnomaly_predictionTimeRange :: Lens' ProactiveAnomaly (Maybe PredictionTimeRange)

Undocumented member.

proactiveAnomaly_limit :: Lens' ProactiveAnomaly (Maybe Double)

A threshold that was exceeded by behavior in analyzed resources. Exceeding this threshold is related to the anomalous behavior that generated this anomaly.

proactiveAnomaly_id :: Lens' ProactiveAnomaly (Maybe Text)

The ID of a proactive anomaly.

proactiveAnomaly_associatedInsightId :: Lens' ProactiveAnomaly (Maybe Text)

The ID of the insight that contains this anomaly. An insight is composed of related anomalies.

proactiveAnomaly_anomalyTimeRange :: Lens' ProactiveAnomaly (Maybe AnomalyTimeRange)

Undocumented member.