Network.AWS.DevOpsGuru.Types.ProactiveInsight

Description

Documentation

data ProactiveInsight

Details about a proactive insight. This object is returned by ListInsights.

See: newProactiveInsight smart constructor.

Constructors

ProactiveInsight'

Fields

status :: Maybe InsightStatus

The status of the proactive insight.

resourceCollection :: Maybe ResourceCollection

severity :: Maybe InsightSeverity

The severity of the proactive insight.

ssmOpsItemId :: Maybe Text

The ID of the AWS System Manager OpsItem created for this insight. You must enable the creation of OpstItems insights before they are created for each insight.

insightTimeRange :: Maybe InsightTimeRange

name :: Maybe Text

The name of the proactive insight.

predictionTimeRange :: Maybe PredictionTimeRange

id :: Maybe Text

The ID of the proactive insight.

newProactiveInsight :: ProactiveInsight

Create a value of ProactiveInsight with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ProactiveInsight', proactiveInsight_status - The status of the proactive insight.

$sel:resourceCollection:ProactiveInsight', proactiveInsight_resourceCollection - Undocumented member.

$sel:severity:ProactiveInsight', proactiveInsight_severity - The severity of the proactive insight.

$sel:ssmOpsItemId:ProactiveInsight', proactiveInsight_ssmOpsItemId - The ID of the AWS System Manager OpsItem created for this insight. You must enable the creation of OpstItems insights before they are created for each insight.

$sel:insightTimeRange:ProactiveInsight', proactiveInsight_insightTimeRange - Undocumented member.

$sel:name:ProactiveInsight', proactiveInsight_name - The name of the proactive insight.

$sel:predictionTimeRange:ProactiveInsight', proactiveInsight_predictionTimeRange - Undocumented member.

$sel:id:ProactiveInsight', proactiveInsight_id - The ID of the proactive insight.

proactiveInsight_status :: Lens' ProactiveInsight (Maybe InsightStatus)

The status of the proactive insight.

proactiveInsight_resourceCollection :: Lens' ProactiveInsight (Maybe ResourceCollection)

Undocumented member.

proactiveInsight_severity :: Lens' ProactiveInsight (Maybe InsightSeverity)

The severity of the proactive insight.

proactiveInsight_ssmOpsItemId :: Lens' ProactiveInsight (Maybe Text)

The ID of the AWS System Manager OpsItem created for this insight. You must enable the creation of OpstItems insights before they are created for each insight.

proactiveInsight_insightTimeRange :: Lens' ProactiveInsight (Maybe InsightTimeRange)

Undocumented member.

proactiveInsight_name :: Lens' ProactiveInsight (Maybe Text)

The name of the proactive insight.

proactiveInsight_predictionTimeRange :: Lens' ProactiveInsight (Maybe PredictionTimeRange)

Undocumented member.

proactiveInsight_id :: Lens' ProactiveInsight (Maybe Text)

The ID of the proactive insight.