Network.AWS.DevOpsGuru.Types.ReactiveAnomaly

Description

Documentation

data ReactiveAnomaly

Details about a reactive anomaly. This object is returned by ListAnomalies.

See: newReactiveAnomaly smart constructor.

Constructors

ReactiveAnomaly'

Fields

anomalyReportedTimeRange :: Maybe AnomalyReportedTimeRange

A AnomalyReportedTimeRange object that specifies the time range between when the anomaly is opened and the time when it is closed.

status :: Maybe AnomalyStatus

The status of the anomaly.

resourceCollection :: Maybe ResourceCollection

severity :: Maybe AnomalySeverity

The severity of the anomaly.

sourceDetails :: Maybe AnomalySourceDetails

Details about the source of the analyzed operational data that triggered the anomaly. The one supported source is Amazon CloudWatch metrics.

id :: Maybe Text

The ID of the reactive anomaly.

associatedInsightId :: Maybe Text

The ID of the insight that contains this anomaly. An insight is composed of related anomalies.

anomalyTimeRange :: Maybe AnomalyTimeRange

newReactiveAnomaly :: ReactiveAnomaly

Create a value of ReactiveAnomaly with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:anomalyReportedTimeRange:ReactiveAnomaly', reactiveAnomaly_anomalyReportedTimeRange - A AnomalyReportedTimeRange object that specifies the time range between when the anomaly is opened and the time when it is closed.

$sel:status:ReactiveAnomaly', reactiveAnomaly_status - The status of the anomaly.

$sel:resourceCollection:ReactiveAnomaly', reactiveAnomaly_resourceCollection - Undocumented member.

$sel:severity:ReactiveAnomaly', reactiveAnomaly_severity - The severity of the anomaly.

$sel:sourceDetails:ReactiveAnomaly', reactiveAnomaly_sourceDetails - Details about the source of the analyzed operational data that triggered the anomaly. The one supported source is Amazon CloudWatch metrics.

$sel:id:ReactiveAnomaly', reactiveAnomaly_id - The ID of the reactive anomaly.

$sel:associatedInsightId:ReactiveAnomaly', reactiveAnomaly_associatedInsightId - The ID of the insight that contains this anomaly. An insight is composed of related anomalies.

$sel:anomalyTimeRange:ReactiveAnomaly', reactiveAnomaly_anomalyTimeRange - Undocumented member.

reactiveAnomaly_anomalyReportedTimeRange :: Lens' ReactiveAnomaly (Maybe AnomalyReportedTimeRange)

A AnomalyReportedTimeRange object that specifies the time range between when the anomaly is opened and the time when it is closed.

reactiveAnomaly_status :: Lens' ReactiveAnomaly (Maybe AnomalyStatus)

The status of the anomaly.

reactiveAnomaly_resourceCollection :: Lens' ReactiveAnomaly (Maybe ResourceCollection)

Undocumented member.

reactiveAnomaly_severity :: Lens' ReactiveAnomaly (Maybe AnomalySeverity)

The severity of the anomaly.

reactiveAnomaly_sourceDetails :: Lens' ReactiveAnomaly (Maybe AnomalySourceDetails)

Details about the source of the analyzed operational data that triggered the anomaly. The one supported source is Amazon CloudWatch metrics.

reactiveAnomaly_id :: Lens' ReactiveAnomaly (Maybe Text)

The ID of the reactive anomaly.

reactiveAnomaly_associatedInsightId :: Lens' ReactiveAnomaly (Maybe Text)

The ID of the insight that contains this anomaly. An insight is composed of related anomalies.

reactiveAnomaly_anomalyTimeRange :: Lens' ReactiveAnomaly (Maybe AnomalyTimeRange)

Undocumented member.