Network.AWS.DevOpsGuru.Types.ServiceResourceCost

Description

Documentation

data ServiceResourceCost

An object that contains information about the estimated monthly cost to analyze an AWS resource. For more information, see Estimate your Amazon DevOps Guru costs and Amazon DevOps Guru pricing.

See: newServiceResourceCost smart constructor.

Constructors

ServiceResourceCost'

Fields

state :: Maybe CostEstimationServiceResourceState

The state of the resource. The resource is ACTIVE if it produces metrics, events, or logs within an hour, otherwise it is INACTIVE. You pay for the number of active AWS resource hours analyzed for each resource. Inactive resources are not charged.

unitCost :: Maybe Double

The price per hour to analyze the resources in the service. For more information, see Estimate your Amazon DevOps Guru costs and Amazon DevOps Guru pricing.

count :: Maybe Int

The number of active resources analyzed for this service to create a monthly cost estimate.

cost :: Maybe Double

The total estimated monthly cost to analyze the active resources for this resource.

type' :: Maybe Text

The type of the AWS resource.

newServiceResourceCost :: ServiceResourceCost

Create a value of ServiceResourceCost with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:ServiceResourceCost', serviceResourceCost_state - The state of the resource. The resource is ACTIVE if it produces metrics, events, or logs within an hour, otherwise it is INACTIVE. You pay for the number of active AWS resource hours analyzed for each resource. Inactive resources are not charged.

$sel:unitCost:ServiceResourceCost', serviceResourceCost_unitCost - The price per hour to analyze the resources in the service. For more information, see Estimate your Amazon DevOps Guru costs and Amazon DevOps Guru pricing.

$sel:count:ServiceResourceCost', serviceResourceCost_count - The number of active resources analyzed for this service to create a monthly cost estimate.

$sel:cost:ServiceResourceCost', serviceResourceCost_cost - The total estimated monthly cost to analyze the active resources for this resource.

$sel:type':ServiceResourceCost', serviceResourceCost_type - The type of the AWS resource.

serviceResourceCost_state :: Lens' ServiceResourceCost (Maybe CostEstimationServiceResourceState)

The state of the resource. The resource is ACTIVE if it produces metrics, events, or logs within an hour, otherwise it is INACTIVE. You pay for the number of active AWS resource hours analyzed for each resource. Inactive resources are not charged.

serviceResourceCost_unitCost :: Lens' ServiceResourceCost (Maybe Double)

The price per hour to analyze the resources in the service. For more information, see Estimate your Amazon DevOps Guru costs and Amazon DevOps Guru pricing.

serviceResourceCost_count :: Lens' ServiceResourceCost (Maybe Int)

The number of active resources analyzed for this service to create a monthly cost estimate.

serviceResourceCost_cost :: Lens' ServiceResourceCost (Maybe Double)

The total estimated monthly cost to analyze the active resources for this resource.

serviceResourceCost_type :: Lens' ServiceResourceCost (Maybe Text)

The type of the AWS resource.