Network.AWS.DevOpsGuru.UpdateServiceIntegration

Description

Enables or disables integration with a service that can be integrated with DevOps Guru. The one service that can be integrated with DevOps Guru is AWS Systems Manager, which can be used to create an OpsItem for each generated insight.

Creating a Request

data UpdateServiceIntegration

See: newUpdateServiceIntegration smart constructor.

Constructors

UpdateServiceIntegration'

Fields

serviceIntegration :: UpdateServiceIntegrationConfig

An IntegratedServiceConfig object used to specify the integrated service you want to update, and whether you want to update it to enabled or disabled.

newUpdateServiceIntegration

Arguments

:: UpdateServiceIntegrationConfig
$sel:serviceIntegration:UpdateServiceIntegration'

-> UpdateServiceIntegration

Create a value of UpdateServiceIntegration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceIntegration:UpdateServiceIntegration', updateServiceIntegration_serviceIntegration - An IntegratedServiceConfig object used to specify the integrated service you want to update, and whether you want to update it to enabled or disabled.

Request Lenses

updateServiceIntegration_serviceIntegration :: Lens' UpdateServiceIntegration UpdateServiceIntegrationConfig

An IntegratedServiceConfig object used to specify the integrated service you want to update, and whether you want to update it to enabled or disabled.

Destructuring the Response

data UpdateServiceIntegrationResponse

See: newUpdateServiceIntegrationResponse smart constructor.

Constructors

UpdateServiceIntegrationResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateServiceIntegrationResponse

Arguments

:: Int
$sel:httpStatus:UpdateServiceIntegrationResponse'

-> UpdateServiceIntegrationResponse

Create a value of UpdateServiceIntegrationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateServiceIntegrationResponse', updateServiceIntegrationResponse_httpStatus - The response's http status code.

Response Lenses

updateServiceIntegrationResponse_httpStatus :: Lens' UpdateServiceIntegrationResponse Int

The response's http status code.