Network.AWS.DeviceFarm.CreateInstanceProfile

Description

Creates a profile that can be applied to one or more private fleet device instances.

Creating a Request

data CreateInstanceProfile

See: newCreateInstanceProfile smart constructor.

Constructors

CreateInstanceProfile'

Fields

excludeAppPackagesFromCleanup :: Maybe [Text]

An array of strings that specifies the list of app packages that should not be cleaned up from the device after a test run.

The list of packages is considered only if you set packageCleanup to true.

description :: Maybe Text

The description of your instance profile.

rebootAfterUse :: Maybe Bool

When set to true, Device Farm reboots the instance after a test run. The default value is true.

packageCleanup :: Maybe Bool

When set to true, Device Farm removes app packages after a test run. The default value is false for private devices.

name :: Text

The name of your instance profile.

newCreateInstanceProfile

Arguments

:: Text
$sel:name:CreateInstanceProfile'

-> CreateInstanceProfile

Create a value of CreateInstanceProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:excludeAppPackagesFromCleanup:CreateInstanceProfile', createInstanceProfile_excludeAppPackagesFromCleanup - An array of strings that specifies the list of app packages that should not be cleaned up from the device after a test run.

The list of packages is considered only if you set packageCleanup to true.

$sel:description:CreateInstanceProfile', createInstanceProfile_description - The description of your instance profile.

$sel:rebootAfterUse:CreateInstanceProfile', createInstanceProfile_rebootAfterUse - When set to true, Device Farm reboots the instance after a test run. The default value is true.

$sel:packageCleanup:CreateInstanceProfile', createInstanceProfile_packageCleanup - When set to true, Device Farm removes app packages after a test run. The default value is false for private devices.

$sel:name:CreateInstanceProfile', createInstanceProfile_name - The name of your instance profile.

Request Lenses

createInstanceProfile_excludeAppPackagesFromCleanup :: Lens' CreateInstanceProfile (Maybe [Text])

An array of strings that specifies the list of app packages that should not be cleaned up from the device after a test run.

The list of packages is considered only if you set packageCleanup to true.

createInstanceProfile_description :: Lens' CreateInstanceProfile (Maybe Text)

The description of your instance profile.

createInstanceProfile_rebootAfterUse :: Lens' CreateInstanceProfile (Maybe Bool)

When set to true, Device Farm reboots the instance after a test run. The default value is true.

createInstanceProfile_packageCleanup :: Lens' CreateInstanceProfile (Maybe Bool)

When set to true, Device Farm removes app packages after a test run. The default value is false for private devices.

createInstanceProfile_name :: Lens' CreateInstanceProfile Text

The name of your instance profile.

Destructuring the Response

data CreateInstanceProfileResponse

See: newCreateInstanceProfileResponse smart constructor.

Constructors

CreateInstanceProfileResponse'

Fields

instanceProfile :: Maybe InstanceProfile

An object that contains information about your instance profile.

httpStatus :: Int

The response's http status code.

newCreateInstanceProfileResponse

Arguments

:: Int
$sel:httpStatus:CreateInstanceProfileResponse'

-> CreateInstanceProfileResponse

Create a value of CreateInstanceProfileResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceProfile:CreateInstanceProfileResponse', createInstanceProfileResponse_instanceProfile - An object that contains information about your instance profile.

$sel:httpStatus:CreateInstanceProfileResponse', createInstanceProfileResponse_httpStatus - The response's http status code.

Response Lenses

createInstanceProfileResponse_instanceProfile :: Lens' CreateInstanceProfileResponse (Maybe InstanceProfile)

An object that contains information about your instance profile.

createInstanceProfileResponse_httpStatus :: Lens' CreateInstanceProfileResponse Int

The response's http status code.