Network.AWS.DeviceFarm.CreateNetworkProfile

Description

Creates a network profile.

Creating a Request

data CreateNetworkProfile

See: newCreateNetworkProfile smart constructor.

Constructors

CreateNetworkProfile'

Fields

uplinkJitterMs :: Maybe Integer

Time variation in the delay of received packets in milliseconds as an integer from 0 to 2000.

uplinkLossPercent :: Maybe Natural

Proportion of transmitted packets that fail to arrive from 0 to 100 percent.

downlinkJitterMs :: Maybe Integer

Time variation in the delay of received packets in milliseconds as an integer from 0 to 2000.

downlinkLossPercent :: Maybe Natural

Proportion of received packets that fail to arrive from 0 to 100 percent.

type' :: Maybe NetworkProfileType

The type of network profile to create. Valid values are listed here.

uplinkDelayMs :: Maybe Integer

Delay time for all packets to destination in milliseconds as an integer from 0 to 2000.

uplinkBandwidthBits :: Maybe Integer

The data throughput rate in bits per second, as an integer from 0 to 104857600.

description :: Maybe Text

The description of the network profile.

downlinkDelayMs :: Maybe Integer

Delay time for all packets to destination in milliseconds as an integer from 0 to 2000.

downlinkBandwidthBits :: Maybe Integer

The data throughput rate in bits per second, as an integer from 0 to 104857600.

projectArn :: Text

The Amazon Resource Name (ARN) of the project for which you want to create a network profile.

name :: Text

The name for the new network profile.

newCreateNetworkProfile

Arguments

:: Text
$sel:projectArn:CreateNetworkProfile'

-> Text
$sel:name:CreateNetworkProfile'

-> CreateNetworkProfile

Create a value of CreateNetworkProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:uplinkJitterMs:CreateNetworkProfile', createNetworkProfile_uplinkJitterMs - Time variation in the delay of received packets in milliseconds as an integer from 0 to 2000.

$sel:uplinkLossPercent:CreateNetworkProfile', createNetworkProfile_uplinkLossPercent - Proportion of transmitted packets that fail to arrive from 0 to 100 percent.

$sel:downlinkJitterMs:CreateNetworkProfile', createNetworkProfile_downlinkJitterMs - Time variation in the delay of received packets in milliseconds as an integer from 0 to 2000.

$sel:downlinkLossPercent:CreateNetworkProfile', createNetworkProfile_downlinkLossPercent - Proportion of received packets that fail to arrive from 0 to 100 percent.

$sel:type':CreateNetworkProfile', createNetworkProfile_type - The type of network profile to create. Valid values are listed here.

$sel:uplinkDelayMs:CreateNetworkProfile', createNetworkProfile_uplinkDelayMs - Delay time for all packets to destination in milliseconds as an integer from 0 to 2000.

$sel:uplinkBandwidthBits:CreateNetworkProfile', createNetworkProfile_uplinkBandwidthBits - The data throughput rate in bits per second, as an integer from 0 to 104857600.

$sel:description:CreateNetworkProfile', createNetworkProfile_description - The description of the network profile.

$sel:downlinkDelayMs:CreateNetworkProfile', createNetworkProfile_downlinkDelayMs - Delay time for all packets to destination in milliseconds as an integer from 0 to 2000.

$sel:downlinkBandwidthBits:CreateNetworkProfile', createNetworkProfile_downlinkBandwidthBits - The data throughput rate in bits per second, as an integer from 0 to 104857600.

$sel:projectArn:CreateNetworkProfile', createNetworkProfile_projectArn - The Amazon Resource Name (ARN) of the project for which you want to create a network profile.

$sel:name:CreateNetworkProfile', createNetworkProfile_name - The name for the new network profile.

Request Lenses

createNetworkProfile_uplinkJitterMs :: Lens' CreateNetworkProfile (Maybe Integer)

Time variation in the delay of received packets in milliseconds as an integer from 0 to 2000.

createNetworkProfile_uplinkLossPercent :: Lens' CreateNetworkProfile (Maybe Natural)

Proportion of transmitted packets that fail to arrive from 0 to 100 percent.

createNetworkProfile_downlinkJitterMs :: Lens' CreateNetworkProfile (Maybe Integer)

Time variation in the delay of received packets in milliseconds as an integer from 0 to 2000.

createNetworkProfile_downlinkLossPercent :: Lens' CreateNetworkProfile (Maybe Natural)

Proportion of received packets that fail to arrive from 0 to 100 percent.

createNetworkProfile_type :: Lens' CreateNetworkProfile (Maybe NetworkProfileType)

The type of network profile to create. Valid values are listed here.

createNetworkProfile_uplinkDelayMs :: Lens' CreateNetworkProfile (Maybe Integer)

Delay time for all packets to destination in milliseconds as an integer from 0 to 2000.

createNetworkProfile_uplinkBandwidthBits :: Lens' CreateNetworkProfile (Maybe Integer)

The data throughput rate in bits per second, as an integer from 0 to 104857600.

createNetworkProfile_description :: Lens' CreateNetworkProfile (Maybe Text)

The description of the network profile.

createNetworkProfile_downlinkDelayMs :: Lens' CreateNetworkProfile (Maybe Integer)

Delay time for all packets to destination in milliseconds as an integer from 0 to 2000.

createNetworkProfile_downlinkBandwidthBits :: Lens' CreateNetworkProfile (Maybe Integer)

The data throughput rate in bits per second, as an integer from 0 to 104857600.

createNetworkProfile_projectArn :: Lens' CreateNetworkProfile Text

The Amazon Resource Name (ARN) of the project for which you want to create a network profile.

createNetworkProfile_name :: Lens' CreateNetworkProfile Text

The name for the new network profile.

Destructuring the Response

data CreateNetworkProfileResponse

See: newCreateNetworkProfileResponse smart constructor.

Constructors

CreateNetworkProfileResponse'

Fields

networkProfile :: Maybe NetworkProfile

The network profile that is returned by the create network profile request.

httpStatus :: Int

The response's http status code.

newCreateNetworkProfileResponse

Arguments

:: Int
$sel:httpStatus:CreateNetworkProfileResponse'

-> CreateNetworkProfileResponse

Create a value of CreateNetworkProfileResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:networkProfile:CreateNetworkProfileResponse', createNetworkProfileResponse_networkProfile - The network profile that is returned by the create network profile request.

$sel:httpStatus:CreateNetworkProfileResponse', createNetworkProfileResponse_httpStatus - The response's http status code.

Response Lenses

createNetworkProfileResponse_networkProfile :: Lens' CreateNetworkProfileResponse (Maybe NetworkProfile)

The network profile that is returned by the create network profile request.

createNetworkProfileResponse_httpStatus :: Lens' CreateNetworkProfileResponse Int

The response's http status code.