Network.AWS.DeviceFarm.CreateRemoteAccessSession

Description

Specifies and starts a remote access session.

Creating a Request

data CreateRemoteAccessSession

Creates and submits a request to start a remote access session.

See: newCreateRemoteAccessSession smart constructor.

Constructors

CreateRemoteAccessSession'

Fields

clientId :: Maybe Text

Unique identifier for the client. If you want access to multiple devices on the same client, you should pass the same clientId value in each call to CreateRemoteAccessSession. This identifier is required only if remoteDebugEnabled is set to true.

Remote debugging is no longer supported.

skipAppResign :: Maybe Bool

When set to true, for private devices, Device Farm does not sign your app again. For public devices, Device Farm always signs your apps again.

For more information on how Device Farm modifies your uploads during tests, see Do you modify my app?

instanceArn :: Maybe Text

The Amazon Resource Name (ARN) of the device instance for which you want to create a remote access session.

remoteRecordEnabled :: Maybe Bool

Set to true to enable remote recording for the remote access session.

remoteRecordAppArn :: Maybe Text

The Amazon Resource Name (ARN) for the app to be recorded in the remote access session.

sshPublicKey :: Maybe Text

Ignored. The public key of the ssh key pair you want to use for connecting to remote devices in your remote debugging session. This key is required only if remoteDebugEnabled is set to true.

Remote debugging is no longer supported.

name :: Maybe Text

The name of the remote access session to create.

remoteDebugEnabled :: Maybe Bool

Set to true if you want to access devices remotely for debugging in your remote access session.

Remote debugging is no longer supported.

configuration :: Maybe CreateRemoteAccessSessionConfiguration

The configuration information for the remote access session request.

interactionMode :: Maybe InteractionMode

The interaction mode of the remote access session. Valid values are:

INTERACTIVE: You can interact with the iOS device by viewing, touching, and rotating the screen. You cannot run XCUITest framework-based tests in this mode.

NO_VIDEO: You are connected to the device, but cannot interact with it or view the screen. This mode has the fastest test execution speed. You can run XCUITest framework-based tests in this mode.

VIDEO_ONLY: You can view the screen, but cannot touch or rotate it. You can run XCUITest framework-based tests and watch the screen in this mode.

projectArn :: Text

The Amazon Resource Name (ARN) of the project for which you want to create a remote access session.

deviceArn :: Text

The ARN of the device for which you want to create a remote access session.

newCreateRemoteAccessSession

Arguments

:: Text
$sel:projectArn:CreateRemoteAccessSession'

-> Text
$sel:deviceArn:CreateRemoteAccessSession'

-> CreateRemoteAccessSession

Create a value of CreateRemoteAccessSession with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientId:CreateRemoteAccessSession', createRemoteAccessSession_clientId - Unique identifier for the client. If you want access to multiple devices on the same client, you should pass the same clientId value in each call to CreateRemoteAccessSession. This identifier is required only if remoteDebugEnabled is set to true.

Remote debugging is no longer supported.

$sel:skipAppResign:CreateRemoteAccessSession', createRemoteAccessSession_skipAppResign - When set to true, for private devices, Device Farm does not sign your app again. For public devices, Device Farm always signs your apps again.

For more information on how Device Farm modifies your uploads during tests, see Do you modify my app?

$sel:instanceArn:CreateRemoteAccessSession', createRemoteAccessSession_instanceArn - The Amazon Resource Name (ARN) of the device instance for which you want to create a remote access session.

$sel:remoteRecordEnabled:CreateRemoteAccessSession', createRemoteAccessSession_remoteRecordEnabled - Set to true to enable remote recording for the remote access session.

$sel:remoteRecordAppArn:CreateRemoteAccessSession', createRemoteAccessSession_remoteRecordAppArn - The Amazon Resource Name (ARN) for the app to be recorded in the remote access session.

$sel:sshPublicKey:CreateRemoteAccessSession', createRemoteAccessSession_sshPublicKey - Ignored. The public key of the ssh key pair you want to use for connecting to remote devices in your remote debugging session. This key is required only if remoteDebugEnabled is set to true.

Remote debugging is no longer supported.

$sel:name:CreateRemoteAccessSession', createRemoteAccessSession_name - The name of the remote access session to create.

$sel:remoteDebugEnabled:CreateRemoteAccessSession', createRemoteAccessSession_remoteDebugEnabled - Set to true if you want to access devices remotely for debugging in your remote access session.

Remote debugging is no longer supported.

$sel:configuration:CreateRemoteAccessSession', createRemoteAccessSession_configuration - The configuration information for the remote access session request.

$sel:interactionMode:CreateRemoteAccessSession', createRemoteAccessSession_interactionMode - The interaction mode of the remote access session. Valid values are:

INTERACTIVE: You can interact with the iOS device by viewing, touching, and rotating the screen. You cannot run XCUITest framework-based tests in this mode.

NO_VIDEO: You are connected to the device, but cannot interact with it or view the screen. This mode has the fastest test execution speed. You can run XCUITest framework-based tests in this mode.

VIDEO_ONLY: You can view the screen, but cannot touch or rotate it. You can run XCUITest framework-based tests and watch the screen in this mode.

$sel:projectArn:CreateRemoteAccessSession', createRemoteAccessSession_projectArn - The Amazon Resource Name (ARN) of the project for which you want to create a remote access session.

$sel:deviceArn:CreateRemoteAccessSession', createRemoteAccessSession_deviceArn - The ARN of the device for which you want to create a remote access session.

Request Lenses

createRemoteAccessSession_clientId :: Lens' CreateRemoteAccessSession (Maybe Text)

Unique identifier for the client. If you want access to multiple devices on the same client, you should pass the same clientId value in each call to CreateRemoteAccessSession. This identifier is required only if remoteDebugEnabled is set to true.

Remote debugging is no longer supported.

createRemoteAccessSession_skipAppResign :: Lens' CreateRemoteAccessSession (Maybe Bool)

When set to true, for private devices, Device Farm does not sign your app again. For public devices, Device Farm always signs your apps again.

For more information on how Device Farm modifies your uploads during tests, see Do you modify my app?

createRemoteAccessSession_instanceArn :: Lens' CreateRemoteAccessSession (Maybe Text)

The Amazon Resource Name (ARN) of the device instance for which you want to create a remote access session.

createRemoteAccessSession_remoteRecordEnabled :: Lens' CreateRemoteAccessSession (Maybe Bool)

Set to true to enable remote recording for the remote access session.

createRemoteAccessSession_remoteRecordAppArn :: Lens' CreateRemoteAccessSession (Maybe Text)

The Amazon Resource Name (ARN) for the app to be recorded in the remote access session.

createRemoteAccessSession_sshPublicKey :: Lens' CreateRemoteAccessSession (Maybe Text)

Ignored. The public key of the ssh key pair you want to use for connecting to remote devices in your remote debugging session. This key is required only if remoteDebugEnabled is set to true.

Remote debugging is no longer supported.

createRemoteAccessSession_name :: Lens' CreateRemoteAccessSession (Maybe Text)

The name of the remote access session to create.

createRemoteAccessSession_remoteDebugEnabled :: Lens' CreateRemoteAccessSession (Maybe Bool)

Set to true if you want to access devices remotely for debugging in your remote access session.

Remote debugging is no longer supported.

createRemoteAccessSession_configuration :: Lens' CreateRemoteAccessSession (Maybe CreateRemoteAccessSessionConfiguration)

The configuration information for the remote access session request.

createRemoteAccessSession_interactionMode :: Lens' CreateRemoteAccessSession (Maybe InteractionMode)

The interaction mode of the remote access session. Valid values are:

INTERACTIVE: You can interact with the iOS device by viewing, touching, and rotating the screen. You cannot run XCUITest framework-based tests in this mode.

NO_VIDEO: You are connected to the device, but cannot interact with it or view the screen. This mode has the fastest test execution speed. You can run XCUITest framework-based tests in this mode.

VIDEO_ONLY: You can view the screen, but cannot touch or rotate it. You can run XCUITest framework-based tests and watch the screen in this mode.

createRemoteAccessSession_projectArn :: Lens' CreateRemoteAccessSession Text

The Amazon Resource Name (ARN) of the project for which you want to create a remote access session.

createRemoteAccessSession_deviceArn :: Lens' CreateRemoteAccessSession Text

The ARN of the device for which you want to create a remote access session.

Destructuring the Response

data CreateRemoteAccessSessionResponse

Represents the server response from a request to create a remote access session.

See: newCreateRemoteAccessSessionResponse smart constructor.

Constructors

CreateRemoteAccessSessionResponse'

Fields

remoteAccessSession :: Maybe RemoteAccessSession

A container that describes the remote access session when the request to create a remote access session is sent.

httpStatus :: Int

The response's http status code.

newCreateRemoteAccessSessionResponse

Arguments

:: Int
$sel:httpStatus:CreateRemoteAccessSessionResponse'

-> CreateRemoteAccessSessionResponse

Create a value of CreateRemoteAccessSessionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:remoteAccessSession:CreateRemoteAccessSessionResponse', createRemoteAccessSessionResponse_remoteAccessSession - A container that describes the remote access session when the request to create a remote access session is sent.

$sel:httpStatus:CreateRemoteAccessSessionResponse', createRemoteAccessSessionResponse_httpStatus - The response's http status code.

Response Lenses

createRemoteAccessSessionResponse_remoteAccessSession :: Lens' CreateRemoteAccessSessionResponse (Maybe RemoteAccessSession)

A container that describes the remote access session when the request to create a remote access session is sent.

createRemoteAccessSessionResponse_httpStatus :: Lens' CreateRemoteAccessSessionResponse Int

The response's http status code.