Network.AWS.DeviceFarm.CreateTestGridUrl

Description

Creates a signed, short-term URL that can be passed to a Selenium RemoteWebDriver constructor.

Creating a Request

data CreateTestGridUrl

See: newCreateTestGridUrl smart constructor.

Constructors

CreateTestGridUrl'

Fields

projectArn :: Text

ARN (from CreateTestGridProject or ListTestGridProjects) to associate with the short-term URL.

expiresInSeconds :: Natural

Lifetime, in seconds, of the URL.

newCreateTestGridUrl

Arguments

:: Text
$sel:projectArn:CreateTestGridUrl'

-> Natural
$sel:expiresInSeconds:CreateTestGridUrl'

-> CreateTestGridUrl

Create a value of CreateTestGridUrl with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:projectArn:CreateTestGridUrl', createTestGridUrl_projectArn - ARN (from CreateTestGridProject or ListTestGridProjects) to associate with the short-term URL.

$sel:expiresInSeconds:CreateTestGridUrl', createTestGridUrl_expiresInSeconds - Lifetime, in seconds, of the URL.

Request Lenses

createTestGridUrl_projectArn :: Lens' CreateTestGridUrl Text

ARN (from CreateTestGridProject or ListTestGridProjects) to associate with the short-term URL.

createTestGridUrl_expiresInSeconds :: Lens' CreateTestGridUrl Natural

Lifetime, in seconds, of the URL.

Destructuring the Response

data CreateTestGridUrlResponse

See: newCreateTestGridUrlResponse smart constructor.

Constructors

CreateTestGridUrlResponse'

Fields

url :: Maybe (Sensitive Text)

A signed URL, expiring in CreateTestGridUrlRequest$expiresInSeconds seconds, to be passed to a RemoteWebDriver.

expires :: Maybe POSIX

The number of seconds the URL from CreateTestGridUrlResult$url stays active.

httpStatus :: Int

The response's http status code.

newCreateTestGridUrlResponse

Arguments

:: Int
$sel:httpStatus:CreateTestGridUrlResponse'

-> CreateTestGridUrlResponse

Create a value of CreateTestGridUrlResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:url:CreateTestGridUrlResponse', createTestGridUrlResponse_url - A signed URL, expiring in CreateTestGridUrlRequest$expiresInSeconds seconds, to be passed to a RemoteWebDriver.

$sel:expires:CreateTestGridUrlResponse', createTestGridUrlResponse_expires - The number of seconds the URL from CreateTestGridUrlResult$url stays active.

$sel:httpStatus:CreateTestGridUrlResponse', createTestGridUrlResponse_httpStatus - The response's http status code.

Response Lenses

createTestGridUrlResponse_url :: Lens' CreateTestGridUrlResponse (Maybe Text)

A signed URL, expiring in CreateTestGridUrlRequest$expiresInSeconds seconds, to be passed to a RemoteWebDriver.

createTestGridUrlResponse_expires :: Lens' CreateTestGridUrlResponse (Maybe UTCTime)

The number of seconds the URL from CreateTestGridUrlResult$url stays active.

createTestGridUrlResponse_httpStatus :: Lens' CreateTestGridUrlResponse Int

The response's http status code.