Network.AWS.DeviceFarm.CreateUpload

Description

Uploads an app or test scripts.

Creating a Request

data CreateUpload

Represents a request to the create upload operation.

See: newCreateUpload smart constructor.

Constructors

CreateUpload'

Fields

contentType :: Maybe Text

The upload's content type (for example, application/octet-stream).

projectArn :: Text

The ARN of the project for the upload.

name :: Text

The upload's file name. The name should not contain any forward slashes (/). If you are uploading an iOS app, the file name must end with the .ipa extension. If you are uploading an Android app, the file name must end with the .apk extension. For all others, the file name must end with the .zip file extension.

type' :: UploadType

The upload's upload type.

Must be one of the following values:

ANDROID_APP

IOS_APP

WEB_APP

EXTERNAL_DATA

APPIUM_JAVA_JUNIT_TEST_PACKAGE

APPIUM_JAVA_TESTNG_TEST_PACKAGE

APPIUM_PYTHON_TEST_PACKAGE

APPIUM_NODE_TEST_PACKAGE

APPIUM_RUBY_TEST_PACKAGE

APPIUM_WEB_JAVA_JUNIT_TEST_PACKAGE

APPIUM_WEB_JAVA_TESTNG_TEST_PACKAGE

APPIUM_WEB_PYTHON_TEST_PACKAGE

APPIUM_WEB_NODE_TEST_PACKAGE

APPIUM_WEB_RUBY_TEST_PACKAGE

CALABASH_TEST_PACKAGE

INSTRUMENTATION_TEST_PACKAGE

UIAUTOMATION_TEST_PACKAGE

UIAUTOMATOR_TEST_PACKAGE

XCTEST_TEST_PACKAGE

XCTEST_UI_TEST_PACKAGE

APPIUM_JAVA_JUNIT_TEST_SPEC

APPIUM_JAVA_TESTNG_TEST_SPEC

APPIUM_PYTHON_TEST_SPEC

APPIUM_NODE_TEST_SPEC

APPIUM_RUBY_TEST_SPEC

APPIUM_WEB_JAVA_JUNIT_TEST_SPEC

APPIUM_WEB_JAVA_TESTNG_TEST_SPEC

APPIUM_WEB_PYTHON_TEST_SPEC

APPIUM_WEB_NODE_TEST_SPEC

APPIUM_WEB_RUBY_TEST_SPEC

INSTRUMENTATION_TEST_SPEC

XCTEST_UI_TEST_SPEC

If you call CreateUpload with WEB_APP specified, AWS Device Farm throws an ArgumentException error.

newCreateUpload

Arguments

:: Text
$sel:projectArn:CreateUpload'

-> Text
$sel:name:CreateUpload'

-> UploadType
$sel:type':CreateUpload'

-> CreateUpload

Create a value of CreateUpload with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:contentType:CreateUpload', createUpload_contentType - The upload's content type (for example, application/octet-stream).

$sel:projectArn:CreateUpload', createUpload_projectArn - The ARN of the project for the upload.

$sel:name:CreateUpload', createUpload_name - The upload's file name. The name should not contain any forward slashes (/). If you are uploading an iOS app, the file name must end with the .ipa extension. If you are uploading an Android app, the file name must end with the .apk extension. For all others, the file name must end with the .zip file extension.

$sel:type':CreateUpload', createUpload_type - The upload's upload type.

Must be one of the following values:

ANDROID_APP

IOS_APP

WEB_APP

EXTERNAL_DATA

APPIUM_JAVA_JUNIT_TEST_PACKAGE

APPIUM_JAVA_TESTNG_TEST_PACKAGE

APPIUM_PYTHON_TEST_PACKAGE

APPIUM_NODE_TEST_PACKAGE

APPIUM_RUBY_TEST_PACKAGE

APPIUM_WEB_JAVA_JUNIT_TEST_PACKAGE

APPIUM_WEB_JAVA_TESTNG_TEST_PACKAGE

APPIUM_WEB_PYTHON_TEST_PACKAGE

APPIUM_WEB_NODE_TEST_PACKAGE

APPIUM_WEB_RUBY_TEST_PACKAGE

CALABASH_TEST_PACKAGE

INSTRUMENTATION_TEST_PACKAGE

UIAUTOMATION_TEST_PACKAGE

UIAUTOMATOR_TEST_PACKAGE

XCTEST_TEST_PACKAGE

XCTEST_UI_TEST_PACKAGE

APPIUM_JAVA_JUNIT_TEST_SPEC

APPIUM_JAVA_TESTNG_TEST_SPEC

APPIUM_PYTHON_TEST_SPEC

APPIUM_NODE_TEST_SPEC

APPIUM_RUBY_TEST_SPEC

APPIUM_WEB_JAVA_JUNIT_TEST_SPEC

APPIUM_WEB_JAVA_TESTNG_TEST_SPEC

APPIUM_WEB_PYTHON_TEST_SPEC

APPIUM_WEB_NODE_TEST_SPEC

APPIUM_WEB_RUBY_TEST_SPEC

INSTRUMENTATION_TEST_SPEC

XCTEST_UI_TEST_SPEC

If you call CreateUpload with WEB_APP specified, AWS Device Farm throws an ArgumentException error.

Request Lenses

createUpload_contentType :: Lens' CreateUpload (Maybe Text)

The upload's content type (for example, application/octet-stream).

createUpload_projectArn :: Lens' CreateUpload Text

The ARN of the project for the upload.

createUpload_name :: Lens' CreateUpload Text

The upload's file name. The name should not contain any forward slashes (/). If you are uploading an iOS app, the file name must end with the .ipa extension. If you are uploading an Android app, the file name must end with the .apk extension. For all others, the file name must end with the .zip file extension.

createUpload_type :: Lens' CreateUpload UploadType

The upload's upload type.

Must be one of the following values:

ANDROID_APP

IOS_APP

WEB_APP

EXTERNAL_DATA

APPIUM_JAVA_JUNIT_TEST_PACKAGE

APPIUM_JAVA_TESTNG_TEST_PACKAGE

APPIUM_PYTHON_TEST_PACKAGE

APPIUM_NODE_TEST_PACKAGE

APPIUM_RUBY_TEST_PACKAGE

APPIUM_WEB_JAVA_JUNIT_TEST_PACKAGE

APPIUM_WEB_JAVA_TESTNG_TEST_PACKAGE

APPIUM_WEB_PYTHON_TEST_PACKAGE

APPIUM_WEB_NODE_TEST_PACKAGE

APPIUM_WEB_RUBY_TEST_PACKAGE

CALABASH_TEST_PACKAGE

INSTRUMENTATION_TEST_PACKAGE

UIAUTOMATION_TEST_PACKAGE

UIAUTOMATOR_TEST_PACKAGE

XCTEST_TEST_PACKAGE

XCTEST_UI_TEST_PACKAGE

APPIUM_JAVA_JUNIT_TEST_SPEC

APPIUM_JAVA_TESTNG_TEST_SPEC

APPIUM_PYTHON_TEST_SPEC

APPIUM_NODE_TEST_SPEC

APPIUM_RUBY_TEST_SPEC

APPIUM_WEB_JAVA_JUNIT_TEST_SPEC

APPIUM_WEB_JAVA_TESTNG_TEST_SPEC

APPIUM_WEB_PYTHON_TEST_SPEC

APPIUM_WEB_NODE_TEST_SPEC

APPIUM_WEB_RUBY_TEST_SPEC

INSTRUMENTATION_TEST_SPEC

XCTEST_UI_TEST_SPEC

If you call CreateUpload with WEB_APP specified, AWS Device Farm throws an ArgumentException error.

Destructuring the Response

data CreateUploadResponse

Represents the result of a create upload request.

See: newCreateUploadResponse smart constructor.

Constructors

CreateUploadResponse'

Fields

upload :: Maybe Upload

The newly created upload.

httpStatus :: Int

The response's http status code.

newCreateUploadResponse

Arguments

:: Int
$sel:httpStatus:CreateUploadResponse'

-> CreateUploadResponse

Create a value of CreateUploadResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:upload:CreateUploadResponse', createUploadResponse_upload - The newly created upload.

$sel:httpStatus:CreateUploadResponse', createUploadResponse_httpStatus - The response's http status code.

Response Lenses

createUploadResponse_upload :: Lens' CreateUploadResponse (Maybe Upload)

The newly created upload.

createUploadResponse_httpStatus :: Lens' CreateUploadResponse Int

The response's http status code.