Network.AWS.DeviceFarm.CreateVPCEConfiguration

Description

Creates a configuration record in Device Farm for your Amazon Virtual Private Cloud (VPC) endpoint.

Creating a Request

data CreateVPCEConfiguration

See: newCreateVPCEConfiguration smart constructor.

Constructors

CreateVPCEConfiguration'

Fields

vpceConfigurationDescription :: Maybe Text

An optional description that provides details about your VPC endpoint configuration.

vpceConfigurationName :: Text

The friendly name you give to your VPC endpoint configuration, to manage your configurations more easily.

vpceServiceName :: Text

The name of the VPC endpoint service running in your AWS account that you want Device Farm to test.

serviceDnsName :: Text

The DNS name of the service running in your VPC that you want Device Farm to test.

newCreateVPCEConfiguration

Arguments

:: Text
$sel:vpceConfigurationName:CreateVPCEConfiguration'

-> Text
$sel:vpceServiceName:CreateVPCEConfiguration'

-> Text
$sel:serviceDnsName:CreateVPCEConfiguration'

-> CreateVPCEConfiguration

Create a value of CreateVPCEConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpceConfigurationDescription:CreateVPCEConfiguration', createVPCEConfiguration_vpceConfigurationDescription - An optional description that provides details about your VPC endpoint configuration.

$sel:vpceConfigurationName:CreateVPCEConfiguration', createVPCEConfiguration_vpceConfigurationName - The friendly name you give to your VPC endpoint configuration, to manage your configurations more easily.

$sel:vpceServiceName:CreateVPCEConfiguration', createVPCEConfiguration_vpceServiceName - The name of the VPC endpoint service running in your AWS account that you want Device Farm to test.

$sel:serviceDnsName:CreateVPCEConfiguration', createVPCEConfiguration_serviceDnsName - The DNS name of the service running in your VPC that you want Device Farm to test.

Request Lenses

createVPCEConfiguration_vpceConfigurationDescription :: Lens' CreateVPCEConfiguration (Maybe Text)

An optional description that provides details about your VPC endpoint configuration.

createVPCEConfiguration_vpceConfigurationName :: Lens' CreateVPCEConfiguration Text

The friendly name you give to your VPC endpoint configuration, to manage your configurations more easily.

createVPCEConfiguration_vpceServiceName :: Lens' CreateVPCEConfiguration Text

The name of the VPC endpoint service running in your AWS account that you want Device Farm to test.

createVPCEConfiguration_serviceDnsName :: Lens' CreateVPCEConfiguration Text

The DNS name of the service running in your VPC that you want Device Farm to test.

Destructuring the Response

data CreateVPCEConfigurationResponse

See: newCreateVPCEConfigurationResponse smart constructor.

Constructors

CreateVPCEConfigurationResponse'

Fields

vpceConfiguration :: Maybe VPCEConfiguration

An object that contains information about your VPC endpoint configuration.

httpStatus :: Int

The response's http status code.

newCreateVPCEConfigurationResponse

Arguments

:: Int
$sel:httpStatus:CreateVPCEConfigurationResponse'

-> CreateVPCEConfigurationResponse

Create a value of CreateVPCEConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpceConfiguration:CreateVPCEConfigurationResponse', createVPCEConfigurationResponse_vpceConfiguration - An object that contains information about your VPC endpoint configuration.

$sel:httpStatus:CreateVPCEConfigurationResponse', createVPCEConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

createVPCEConfigurationResponse_vpceConfiguration :: Lens' CreateVPCEConfigurationResponse (Maybe VPCEConfiguration)

An object that contains information about your VPC endpoint configuration.

createVPCEConfigurationResponse_httpStatus :: Lens' CreateVPCEConfigurationResponse Int

The response's http status code.