Network.AWS.DeviceFarm.GetRun

Description

Gets information about a run.

Creating a Request

data GetRun

Represents a request to the get run operation.

See: newGetRun smart constructor.

Constructors

GetRun'

Fields

arn :: Text

The run's ARN.

newGetRun

Arguments

:: Text
$sel:arn:GetRun'

-> GetRun

Create a value of GetRun with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:GetRun', getRun_arn - The run's ARN.

Request Lenses

getRun_arn :: Lens' GetRun Text

The run's ARN.

Destructuring the Response

data GetRunResponse

Represents the result of a get run request.

See: newGetRunResponse smart constructor.

Constructors

GetRunResponse'

Fields

run :: Maybe Run

The run to get results from.

httpStatus :: Int

The response's http status code.

newGetRunResponse

Arguments

:: Int
$sel:httpStatus:GetRunResponse'

-> GetRunResponse

Create a value of GetRunResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:run:GetRunResponse', getRunResponse_run - The run to get results from.

$sel:httpStatus:GetRunResponse', getRunResponse_httpStatus - The response's http status code.

Response Lenses

getRunResponse_run :: Lens' GetRunResponse (Maybe Run)

The run to get results from.

getRunResponse_httpStatus :: Lens' GetRunResponse Int

The response's http status code.