Network.AWS.DeviceFarm.GetTest

Description

Gets information about a test.

Creating a Request

data GetTest

Represents a request to the get test operation.

See: newGetTest smart constructor.

Constructors

GetTest'

Fields

arn :: Text

The test's ARN.

newGetTest

Arguments

:: Text
$sel:arn:GetTest'

-> GetTest

Create a value of GetTest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:GetTest', getTest_arn - The test's ARN.

Request Lenses

getTest_arn :: Lens' GetTest Text

The test's ARN.

Destructuring the Response

data GetTestResponse

Represents the result of a get test request.

See: newGetTestResponse smart constructor.

Constructors

GetTestResponse'

Fields

test :: Maybe Test

A test condition that is evaluated.

httpStatus :: Int

The response's http status code.

newGetTestResponse

Arguments

:: Int
$sel:httpStatus:GetTestResponse'

-> GetTestResponse

Create a value of GetTestResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:test:GetTestResponse', getTestResponse_test - A test condition that is evaluated.

$sel:httpStatus:GetTestResponse', getTestResponse_httpStatus - The response's http status code.

Response Lenses

getTestResponse_test :: Lens' GetTestResponse (Maybe Test)

A test condition that is evaluated.

getTestResponse_httpStatus :: Lens' GetTestResponse Int

The response's http status code.