Network.AWS.DeviceFarm.GetTestGridProject

Description

Retrieves information about a Selenium testing project.

Creating a Request

data GetTestGridProject

See: newGetTestGridProject smart constructor.

Constructors

GetTestGridProject'

Fields

projectArn :: Text

The ARN of the Selenium testing project, from either CreateTestGridProject or ListTestGridProjects.

newGetTestGridProject

Arguments

:: Text
$sel:projectArn:GetTestGridProject'

-> GetTestGridProject

Create a value of GetTestGridProject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:projectArn:GetTestGridProject', getTestGridProject_projectArn - The ARN of the Selenium testing project, from either CreateTestGridProject or ListTestGridProjects.

Request Lenses

getTestGridProject_projectArn :: Lens' GetTestGridProject Text

The ARN of the Selenium testing project, from either CreateTestGridProject or ListTestGridProjects.

Destructuring the Response

data GetTestGridProjectResponse

See: newGetTestGridProjectResponse smart constructor.

Constructors

GetTestGridProjectResponse'

Fields

testGridProject :: Maybe TestGridProject

A TestGridProject.

httpStatus :: Int

The response's http status code.

newGetTestGridProjectResponse

Arguments

:: Int
$sel:httpStatus:GetTestGridProjectResponse'

-> GetTestGridProjectResponse

Create a value of GetTestGridProjectResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:testGridProject:GetTestGridProjectResponse', getTestGridProjectResponse_testGridProject - A TestGridProject.

$sel:httpStatus:GetTestGridProjectResponse', getTestGridProjectResponse_httpStatus - The response's http status code.

Response Lenses

getTestGridProjectResponse_testGridProject :: Lens' GetTestGridProjectResponse (Maybe TestGridProject)

A TestGridProject.

getTestGridProjectResponse_httpStatus :: Lens' GetTestGridProjectResponse Int

The response's http status code.