Network.AWS.DeviceFarm.InstallToRemoteAccessSession

Description

Installs an application to the device in a remote access session. For Android applications, the file must be in .apk format. For iOS applications, the file must be in .ipa format.

Creating a Request

data InstallToRemoteAccessSession

Represents the request to install an Android application (in .apk format) or an iOS application (in .ipa format) as part of a remote access session.

See: newInstallToRemoteAccessSession smart constructor.

Constructors

InstallToRemoteAccessSession'

Fields

remoteAccessSessionArn :: Text

The Amazon Resource Name (ARN) of the remote access session about which you are requesting information.

appArn :: Text

The ARN of the app about which you are requesting information.

newInstallToRemoteAccessSession

Arguments

:: Text
$sel:remoteAccessSessionArn:InstallToRemoteAccessSession'

-> Text
$sel:appArn:InstallToRemoteAccessSession'

-> InstallToRemoteAccessSession

Create a value of InstallToRemoteAccessSession with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:remoteAccessSessionArn:InstallToRemoteAccessSession', installToRemoteAccessSession_remoteAccessSessionArn - The Amazon Resource Name (ARN) of the remote access session about which you are requesting information.

$sel:appArn:InstallToRemoteAccessSession', installToRemoteAccessSession_appArn - The ARN of the app about which you are requesting information.

Request Lenses

installToRemoteAccessSession_remoteAccessSessionArn :: Lens' InstallToRemoteAccessSession Text

The Amazon Resource Name (ARN) of the remote access session about which you are requesting information.

installToRemoteAccessSession_appArn :: Lens' InstallToRemoteAccessSession Text

The ARN of the app about which you are requesting information.

Destructuring the Response

data InstallToRemoteAccessSessionResponse

Represents the response from the server after AWS Device Farm makes a request to install to a remote access session.

See: newInstallToRemoteAccessSessionResponse smart constructor.

Constructors

InstallToRemoteAccessSessionResponse'

Fields

appUpload :: Maybe Upload

An app to upload or that has been uploaded.

httpStatus :: Int

The response's http status code.

newInstallToRemoteAccessSessionResponse

Arguments

:: Int
$sel:httpStatus:InstallToRemoteAccessSessionResponse'

-> InstallToRemoteAccessSessionResponse

Create a value of InstallToRemoteAccessSessionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appUpload:InstallToRemoteAccessSessionResponse', installToRemoteAccessSessionResponse_appUpload - An app to upload or that has been uploaded.

$sel:httpStatus:InstallToRemoteAccessSessionResponse', installToRemoteAccessSessionResponse_httpStatus - The response's http status code.

Response Lenses

installToRemoteAccessSessionResponse_appUpload :: Lens' InstallToRemoteAccessSessionResponse (Maybe Upload)

An app to upload or that has been uploaded.

installToRemoteAccessSessionResponse_httpStatus :: Lens' InstallToRemoteAccessSessionResponse Int

The response's http status code.