Network.AWS.DeviceFarm.ListArtifacts

Description

Gets information about artifacts.

This operation returns paginated results.

Creating a Request

data ListArtifacts

Represents a request to the list artifacts operation.

See: newListArtifacts smart constructor.

Constructors

ListArtifacts'

Fields

nextToken :: Maybe Text

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

arn :: Text

The run, job, suite, or test ARN.

type' :: ArtifactCategory

The artifacts' type.

Allowed values include:

FILE

LOG

SCREENSHOT

newListArtifacts

Arguments

:: Text
$sel:arn:ListArtifacts'

-> ArtifactCategory
$sel:type':ListArtifacts'

-> ListArtifacts

Create a value of ListArtifacts with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListArtifacts', listArtifacts_nextToken - An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

$sel:arn:ListArtifacts', listArtifacts_arn - The run, job, suite, or test ARN.

$sel:type':ListArtifacts', listArtifacts_type - The artifacts' type.

Allowed values include:

FILE

LOG

SCREENSHOT

Request Lenses

listArtifacts_nextToken :: Lens' ListArtifacts (Maybe Text)

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

listArtifacts_arn :: Lens' ListArtifacts Text

The run, job, suite, or test ARN.

listArtifacts_type :: Lens' ListArtifacts ArtifactCategory

The artifacts' type.

Allowed values include:

FILE

LOG

SCREENSHOT

Destructuring the Response

data ListArtifactsResponse

Represents the result of a list artifacts operation.

See: newListArtifactsResponse smart constructor.

Constructors

ListArtifactsResponse'

Fields

nextToken :: Maybe Text

If the number of items that are returned is significantly large, this is an identifier that is also returned. It can be used in a subsequent call to this operation to return the next set of items in the list.

artifacts :: Maybe [Artifact]

Information about the artifacts.

httpStatus :: Int

The response's http status code.

newListArtifactsResponse

Arguments

:: Int
$sel:httpStatus:ListArtifactsResponse'

-> ListArtifactsResponse

Create a value of ListArtifactsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListArtifacts', listArtifactsResponse_nextToken - If the number of items that are returned is significantly large, this is an identifier that is also returned. It can be used in a subsequent call to this operation to return the next set of items in the list.

$sel:artifacts:ListArtifactsResponse', listArtifactsResponse_artifacts - Information about the artifacts.

$sel:httpStatus:ListArtifactsResponse', listArtifactsResponse_httpStatus - The response's http status code.

Response Lenses

listArtifactsResponse_nextToken :: Lens' ListArtifactsResponse (Maybe Text)

If the number of items that are returned is significantly large, this is an identifier that is also returned. It can be used in a subsequent call to this operation to return the next set of items in the list.

listArtifactsResponse_artifacts :: Lens' ListArtifactsResponse (Maybe [Artifact])

Information about the artifacts.

listArtifactsResponse_httpStatus :: Lens' ListArtifactsResponse Int

The response's http status code.