Network.AWS.DeviceFarm.ListDevicePools

Description

Gets information about device pools.

This operation returns paginated results.

Creating a Request

data ListDevicePools

Represents the result of a list device pools request.

See: newListDevicePools smart constructor.

Constructors

ListDevicePools'

Fields

nextToken :: Maybe Text

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

type' :: Maybe DevicePoolType

The device pools' type.

Allowed values include:

CURATED: A device pool that is created and managed by AWS Device Farm.

PRIVATE: A device pool that is created and managed by the device pool developer.

arn :: Text

The project ARN.

newListDevicePools

Arguments

:: Text
$sel:arn:ListDevicePools'

-> ListDevicePools

Create a value of ListDevicePools with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDevicePools', listDevicePools_nextToken - An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

$sel:type':ListDevicePools', listDevicePools_type - The device pools' type.

Allowed values include:

CURATED: A device pool that is created and managed by AWS Device Farm.

PRIVATE: A device pool that is created and managed by the device pool developer.

$sel:arn:ListDevicePools', listDevicePools_arn - The project ARN.

Request Lenses

listDevicePools_nextToken :: Lens' ListDevicePools (Maybe Text)

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

listDevicePools_type :: Lens' ListDevicePools (Maybe DevicePoolType)

The device pools' type.

Allowed values include:

CURATED: A device pool that is created and managed by AWS Device Farm.

PRIVATE: A device pool that is created and managed by the device pool developer.

listDevicePools_arn :: Lens' ListDevicePools Text

The project ARN.

Destructuring the Response

data ListDevicePoolsResponse

Represents the result of a list device pools request.

See: newListDevicePoolsResponse smart constructor.

Constructors

ListDevicePoolsResponse'

Fields

nextToken :: Maybe Text

If the number of items that are returned is significantly large, this is an identifier that is also returned. It can be used in a subsequent call to this operation to return the next set of items in the list.

devicePools :: Maybe [DevicePool]

Information about the device pools.

httpStatus :: Int

The response's http status code.

newListDevicePoolsResponse

Arguments

:: Int
$sel:httpStatus:ListDevicePoolsResponse'

-> ListDevicePoolsResponse

Create a value of ListDevicePoolsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDevicePools', listDevicePoolsResponse_nextToken - If the number of items that are returned is significantly large, this is an identifier that is also returned. It can be used in a subsequent call to this operation to return the next set of items in the list.

$sel:devicePools:ListDevicePoolsResponse', listDevicePoolsResponse_devicePools - Information about the device pools.

$sel:httpStatus:ListDevicePoolsResponse', listDevicePoolsResponse_httpStatus - The response's http status code.

Response Lenses

listDevicePoolsResponse_nextToken :: Lens' ListDevicePoolsResponse (Maybe Text)

If the number of items that are returned is significantly large, this is an identifier that is also returned. It can be used in a subsequent call to this operation to return the next set of items in the list.

listDevicePoolsResponse_devicePools :: Lens' ListDevicePoolsResponse (Maybe [DevicePool])

Information about the device pools.

listDevicePoolsResponse_httpStatus :: Lens' ListDevicePoolsResponse Int

The response's http status code.