Network.AWS.DeviceFarm.ListDevices

Description

Gets information about unique device types.

This operation returns paginated results.

Creating a Request

data ListDevices

Represents the result of a list devices request.

See: newListDevices smart constructor.

Constructors

ListDevices'

Fields

arn :: Maybe Text

The Amazon Resource Name (ARN) of the project.

filters :: Maybe [DeviceFilter]

Used to select a set of devices. A filter is made up of an attribute, an operator, and one or more values.

Attribute: The aspect of a device such as platform or model used as the selection criteria in a device filter.

Allowed values include:

ARN: The Amazon Resource Name (ARN) of the device (for example, arn:aws:devicefarm:us-west-2::device:12345Example).

PLATFORM: The device platform. Valid values are ANDROID or IOS.

OS_VERSION: The operating system version (for example, 10.3.2).

MODEL: The device model (for example, iPad 5th Gen).

AVAILABILITY: The current availability of the device. Valid values are AVAILABLE, HIGHLY_AVAILABLE, BUSY, or TEMPORARY_NOT_AVAILABLE.

FORM_FACTOR: The device form factor. Valid values are PHONE or TABLET.

MANUFACTURER: The device manufacturer (for example, Apple).

REMOTE_ACCESS_ENABLED: Whether the device is enabled for remote access. Valid values are TRUE or FALSE.

REMOTE_DEBUG_ENABLED: Whether the device is enabled for remote debugging. Valid values are TRUE or FALSE. Because remote debugging is no longer supported, this attribute is ignored.

INSTANCE_ARN: The Amazon Resource Name (ARN) of the device instance.

INSTANCE_LABELS: The label of the device instance.

FLEET_TYPE: The fleet type. Valid values are PUBLIC or PRIVATE.

Operator: The filter operator.

The EQUALS operator is available for every attribute except INSTANCE_LABELS.

The CONTAINS operator is available for the INSTANCE_LABELS and MODEL attributes.

The IN and NOT_IN operators are available for the ARN, OS_VERSION, MODEL, MANUFACTURER, and INSTANCE_ARN attributes.

The LESS_THAN, GREATER_THAN, LESS_THAN_OR_EQUALS, and GREATER_THAN_OR_EQUALS operators are also available for the OS_VERSION attribute.

Values: An array of one or more filter values.

The IN and NOT_IN operators take a values array that has one or more elements.

The other operators require an array with a single element.

In a request, the AVAILABILITY attribute takes the following values: AVAILABLE, HIGHLY_AVAILABLE, BUSY, or TEMPORARY_NOT_AVAILABLE.

nextToken :: Maybe Text

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

newListDevices :: ListDevices

Create a value of ListDevices with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:ListDevices', listDevices_arn - The Amazon Resource Name (ARN) of the project.

$sel:filters:ListDevices', listDevices_filters - Used to select a set of devices. A filter is made up of an attribute, an operator, and one or more values.

Attribute: The aspect of a device such as platform or model used as the selection criteria in a device filter.

Allowed values include:

ARN: The Amazon Resource Name (ARN) of the device (for example, arn:aws:devicefarm:us-west-2::device:12345Example).

PLATFORM: The device platform. Valid values are ANDROID or IOS.

OS_VERSION: The operating system version (for example, 10.3.2).

MODEL: The device model (for example, iPad 5th Gen).

AVAILABILITY: The current availability of the device. Valid values are AVAILABLE, HIGHLY_AVAILABLE, BUSY, or TEMPORARY_NOT_AVAILABLE.

FORM_FACTOR: The device form factor. Valid values are PHONE or TABLET.

MANUFACTURER: The device manufacturer (for example, Apple).

REMOTE_ACCESS_ENABLED: Whether the device is enabled for remote access. Valid values are TRUE or FALSE.

REMOTE_DEBUG_ENABLED: Whether the device is enabled for remote debugging. Valid values are TRUE or FALSE. Because remote debugging is no longer supported, this attribute is ignored.

INSTANCE_ARN: The Amazon Resource Name (ARN) of the device instance.

INSTANCE_LABELS: The label of the device instance.

FLEET_TYPE: The fleet type. Valid values are PUBLIC or PRIVATE.

Operator: The filter operator.

The EQUALS operator is available for every attribute except INSTANCE_LABELS.

The CONTAINS operator is available for the INSTANCE_LABELS and MODEL attributes.

The IN and NOT_IN operators are available for the ARN, OS_VERSION, MODEL, MANUFACTURER, and INSTANCE_ARN attributes.

The LESS_THAN, GREATER_THAN, LESS_THAN_OR_EQUALS, and GREATER_THAN_OR_EQUALS operators are also available for the OS_VERSION attribute.

Values: An array of one or more filter values.

The IN and NOT_IN operators take a values array that has one or more elements.

The other operators require an array with a single element.

In a request, the AVAILABILITY attribute takes the following values: AVAILABLE, HIGHLY_AVAILABLE, BUSY, or TEMPORARY_NOT_AVAILABLE.

$sel:nextToken:ListDevices', listDevices_nextToken - An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Request Lenses

listDevices_arn :: Lens' ListDevices (Maybe Text)

The Amazon Resource Name (ARN) of the project.

listDevices_filters :: Lens' ListDevices (Maybe [DeviceFilter])

Used to select a set of devices. A filter is made up of an attribute, an operator, and one or more values.

Attribute: The aspect of a device such as platform or model used as the selection criteria in a device filter.

Allowed values include:

ARN: The Amazon Resource Name (ARN) of the device (for example, arn:aws:devicefarm:us-west-2::device:12345Example).

PLATFORM: The device platform. Valid values are ANDROID or IOS.

OS_VERSION: The operating system version (for example, 10.3.2).

MODEL: The device model (for example, iPad 5th Gen).

AVAILABILITY: The current availability of the device. Valid values are AVAILABLE, HIGHLY_AVAILABLE, BUSY, or TEMPORARY_NOT_AVAILABLE.

FORM_FACTOR: The device form factor. Valid values are PHONE or TABLET.

MANUFACTURER: The device manufacturer (for example, Apple).

REMOTE_ACCESS_ENABLED: Whether the device is enabled for remote access. Valid values are TRUE or FALSE.

REMOTE_DEBUG_ENABLED: Whether the device is enabled for remote debugging. Valid values are TRUE or FALSE. Because remote debugging is no longer supported, this attribute is ignored.

INSTANCE_ARN: The Amazon Resource Name (ARN) of the device instance.

INSTANCE_LABELS: The label of the device instance.

FLEET_TYPE: The fleet type. Valid values are PUBLIC or PRIVATE.

Operator: The filter operator.

The EQUALS operator is available for every attribute except INSTANCE_LABELS.

The CONTAINS operator is available for the INSTANCE_LABELS and MODEL attributes.

The IN and NOT_IN operators are available for the ARN, OS_VERSION, MODEL, MANUFACTURER, and INSTANCE_ARN attributes.

The LESS_THAN, GREATER_THAN, LESS_THAN_OR_EQUALS, and GREATER_THAN_OR_EQUALS operators are also available for the OS_VERSION attribute.

Values: An array of one or more filter values.

The IN and NOT_IN operators take a values array that has one or more elements.

The other operators require an array with a single element.

In a request, the AVAILABILITY attribute takes the following values: AVAILABLE, HIGHLY_AVAILABLE, BUSY, or TEMPORARY_NOT_AVAILABLE.

listDevices_nextToken :: Lens' ListDevices (Maybe Text)

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Destructuring the Response

data ListDevicesResponse

Represents the result of a list devices operation.

See: newListDevicesResponse smart constructor.

Constructors

ListDevicesResponse'

Fields

nextToken :: Maybe Text

If the number of items that are returned is significantly large, this is an identifier that is also returned. It can be used in a subsequent call to this operation to return the next set of items in the list.

devices :: Maybe [Device]

Information about the devices.

httpStatus :: Int

The response's http status code.

newListDevicesResponse

Arguments

:: Int
$sel:httpStatus:ListDevicesResponse'

-> ListDevicesResponse

Create a value of ListDevicesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDevices', listDevicesResponse_nextToken - If the number of items that are returned is significantly large, this is an identifier that is also returned. It can be used in a subsequent call to this operation to return the next set of items in the list.

$sel:devices:ListDevicesResponse', listDevicesResponse_devices - Information about the devices.

$sel:httpStatus:ListDevicesResponse', listDevicesResponse_httpStatus - The response's http status code.

Response Lenses

listDevicesResponse_nextToken :: Lens' ListDevicesResponse (Maybe Text)

If the number of items that are returned is significantly large, this is an identifier that is also returned. It can be used in a subsequent call to this operation to return the next set of items in the list.

listDevicesResponse_devices :: Lens' ListDevicesResponse (Maybe [Device])

Information about the devices.

listDevicesResponse_httpStatus :: Lens' ListDevicesResponse Int

The response's http status code.