Network.AWS.DeviceFarm.ListOfferings

Description

Returns a list of products or offerings that the user can manage through the API. Each offering record indicates the recurring price per unit and the frequency for that offering. The API returns a NotEligible error if the user is not permitted to invoke the operation. If you must be able to invoke this operation, contact aws-devicefarm-support@amazon.com.

This operation returns paginated results.

Creating a Request

data ListOfferings

Represents the request to list all offerings.

See: newListOfferings smart constructor.

Constructors

ListOfferings'

Fields

nextToken :: Maybe Text

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

newListOfferings :: ListOfferings

Create a value of ListOfferings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListOfferings', listOfferings_nextToken - An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Request Lenses

listOfferings_nextToken :: Lens' ListOfferings (Maybe Text)

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Destructuring the Response

data ListOfferingsResponse

Represents the return values of the list of offerings.

See: newListOfferingsResponse smart constructor.

Constructors

ListOfferingsResponse'

Fields

nextToken :: Maybe Text

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

offerings :: Maybe [Offering]

A value that represents the list offering results.

httpStatus :: Int

The response's http status code.

newListOfferingsResponse

Arguments

:: Int
$sel:httpStatus:ListOfferingsResponse'

-> ListOfferingsResponse

Create a value of ListOfferingsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListOfferings', listOfferingsResponse_nextToken - An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

$sel:offerings:ListOfferingsResponse', listOfferingsResponse_offerings - A value that represents the list offering results.

$sel:httpStatus:ListOfferingsResponse', listOfferingsResponse_httpStatus - The response's http status code.

Response Lenses

listOfferingsResponse_nextToken :: Lens' ListOfferingsResponse (Maybe Text)

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

listOfferingsResponse_offerings :: Lens' ListOfferingsResponse (Maybe [Offering])

A value that represents the list offering results.

listOfferingsResponse_httpStatus :: Lens' ListOfferingsResponse Int

The response's http status code.