Network.AWS.DeviceFarm.ListTestGridSessions

Description

Retrieves a list of sessions for a TestGridProject.

Creating a Request

data ListTestGridSessions

See: newListTestGridSessions smart constructor.

Constructors

ListTestGridSessions'

Fields

nextToken :: Maybe Text

Pagination token.

status :: Maybe TestGridSessionStatus

Return only sessions in this state.

creationTimeBefore :: Maybe POSIX

Return only sessions created before this time.

maxResult :: Maybe Natural

Return only this many results at a time.

endTimeAfter :: Maybe POSIX

Return only sessions that ended after this time.

endTimeBefore :: Maybe POSIX

Return only sessions that ended before this time.

creationTimeAfter :: Maybe POSIX

Return only sessions created after this time.

projectArn :: Text

ARN of a TestGridProject.

newListTestGridSessions

Arguments

:: Text
$sel:projectArn:ListTestGridSessions'

-> ListTestGridSessions

Create a value of ListTestGridSessions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListTestGridSessions', listTestGridSessions_nextToken - Pagination token.

$sel:status:ListTestGridSessions', listTestGridSessions_status - Return only sessions in this state.

$sel:creationTimeBefore:ListTestGridSessions', listTestGridSessions_creationTimeBefore - Return only sessions created before this time.

$sel:maxResult:ListTestGridSessions', listTestGridSessions_maxResult - Return only this many results at a time.

$sel:endTimeAfter:ListTestGridSessions', listTestGridSessions_endTimeAfter - Return only sessions that ended after this time.

$sel:endTimeBefore:ListTestGridSessions', listTestGridSessions_endTimeBefore - Return only sessions that ended before this time.

$sel:creationTimeAfter:ListTestGridSessions', listTestGridSessions_creationTimeAfter - Return only sessions created after this time.

$sel:projectArn:ListTestGridSessions', listTestGridSessions_projectArn - ARN of a TestGridProject.

Request Lenses

listTestGridSessions_nextToken :: Lens' ListTestGridSessions (Maybe Text)

Pagination token.

listTestGridSessions_status :: Lens' ListTestGridSessions (Maybe TestGridSessionStatus)

Return only sessions in this state.

listTestGridSessions_creationTimeBefore :: Lens' ListTestGridSessions (Maybe UTCTime)

Return only sessions created before this time.

listTestGridSessions_maxResult :: Lens' ListTestGridSessions (Maybe Natural)

Return only this many results at a time.

listTestGridSessions_endTimeAfter :: Lens' ListTestGridSessions (Maybe UTCTime)

Return only sessions that ended after this time.

listTestGridSessions_endTimeBefore :: Lens' ListTestGridSessions (Maybe UTCTime)

Return only sessions that ended before this time.

listTestGridSessions_creationTimeAfter :: Lens' ListTestGridSessions (Maybe UTCTime)

Return only sessions created after this time.

listTestGridSessions_projectArn :: Lens' ListTestGridSessions Text

ARN of a TestGridProject.

Destructuring the Response

data ListTestGridSessionsResponse

See: newListTestGridSessionsResponse smart constructor.

Constructors

ListTestGridSessionsResponse'

Fields

nextToken :: Maybe Text

Pagination token.

testGridSessions :: Maybe [TestGridSession]

The sessions that match the criteria in a ListTestGridSessionsRequest.

httpStatus :: Int

The response's http status code.

newListTestGridSessionsResponse

Arguments

:: Int
$sel:httpStatus:ListTestGridSessionsResponse'

-> ListTestGridSessionsResponse

Create a value of ListTestGridSessionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListTestGridSessions', listTestGridSessionsResponse_nextToken - Pagination token.

$sel:testGridSessions:ListTestGridSessionsResponse', listTestGridSessionsResponse_testGridSessions - The sessions that match the criteria in a ListTestGridSessionsRequest.

$sel:httpStatus:ListTestGridSessionsResponse', listTestGridSessionsResponse_httpStatus - The response's http status code.

Response Lenses

listTestGridSessionsResponse_nextToken :: Lens' ListTestGridSessionsResponse (Maybe Text)

Pagination token.

listTestGridSessionsResponse_testGridSessions :: Lens' ListTestGridSessionsResponse (Maybe [TestGridSession])

The sessions that match the criteria in a ListTestGridSessionsRequest.

listTestGridSessionsResponse_httpStatus :: Lens' ListTestGridSessionsResponse Int

The response's http status code.