Network.AWS.DeviceFarm.RenewOffering

Description

Explicitly sets the quantity of devices to renew for an offering, starting from the effectiveDate of the next period. The API returns a NotEligible error if the user is not permitted to invoke the operation. If you must be able to invoke this operation, contact aws-devicefarm-support@amazon.com.

Creating a Request

data RenewOffering

A request that represents an offering renewal.

See: newRenewOffering smart constructor.

Constructors

RenewOffering'

Fields

offeringId :: Text

The ID of a request to renew an offering.

quantity :: Int

The quantity requested in an offering renewal.

newRenewOffering

Arguments

:: Text
$sel:offeringId:RenewOffering'

-> Int
$sel:quantity:RenewOffering'

-> RenewOffering

Create a value of RenewOffering with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:offeringId:RenewOffering', renewOffering_offeringId - The ID of a request to renew an offering.

$sel:quantity:RenewOffering', renewOffering_quantity - The quantity requested in an offering renewal.

Request Lenses

renewOffering_offeringId :: Lens' RenewOffering Text

The ID of a request to renew an offering.

renewOffering_quantity :: Lens' RenewOffering Int

The quantity requested in an offering renewal.

Destructuring the Response

data RenewOfferingResponse

The result of a renewal offering.

See: newRenewOfferingResponse smart constructor.

Constructors

RenewOfferingResponse'

Fields

offeringTransaction :: Maybe OfferingTransaction

Represents the status of the offering transaction for the renewal.

httpStatus :: Int

The response's http status code.

newRenewOfferingResponse

Arguments

:: Int
$sel:httpStatus:RenewOfferingResponse'

-> RenewOfferingResponse

Create a value of RenewOfferingResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:offeringTransaction:RenewOfferingResponse', renewOfferingResponse_offeringTransaction - Represents the status of the offering transaction for the renewal.

$sel:httpStatus:RenewOfferingResponse', renewOfferingResponse_httpStatus - The response's http status code.

Response Lenses

renewOfferingResponse_offeringTransaction :: Lens' RenewOfferingResponse (Maybe OfferingTransaction)

Represents the status of the offering transaction for the renewal.

renewOfferingResponse_httpStatus :: Lens' RenewOfferingResponse Int

The response's http status code.