Network.AWS.DeviceFarm.StopRun

Description

Initiates a stop request for the current test run. AWS Device Farm immediately stops the run on devices where tests have not started. You are not billed for these devices. On devices where tests have started executing, setup suite and teardown suite tests run to completion on those devices. You are billed for setup, teardown, and any tests that were in progress or already completed.

Creating a Request

data StopRun

Represents the request to stop a specific run.

See: newStopRun smart constructor.

Constructors

StopRun'

Fields

arn :: Text

Represents the Amazon Resource Name (ARN) of the Device Farm run to stop.

newStopRun

Arguments

:: Text
$sel:arn:StopRun'

-> StopRun

Create a value of StopRun with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:StopRun', stopRun_arn - Represents the Amazon Resource Name (ARN) of the Device Farm run to stop.

Request Lenses

stopRun_arn :: Lens' StopRun Text

Represents the Amazon Resource Name (ARN) of the Device Farm run to stop.

Destructuring the Response

data StopRunResponse

Represents the results of your stop run attempt.

See: newStopRunResponse smart constructor.

Constructors

StopRunResponse'

Fields

run :: Maybe Run

The run that was stopped.

httpStatus :: Int

The response's http status code.

newStopRunResponse

Arguments

:: Int
$sel:httpStatus:StopRunResponse'

-> StopRunResponse

Create a value of StopRunResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:run:StopRunResponse', stopRunResponse_run - The run that was stopped.

$sel:httpStatus:StopRunResponse', stopRunResponse_httpStatus - The response's http status code.

Response Lenses

stopRunResponse_run :: Lens' StopRunResponse (Maybe Run)

The run that was stopped.

stopRunResponse_httpStatus :: Lens' StopRunResponse Int

The response's http status code.