Network.AWS.DeviceFarm.Types.AccountSettings

Description

Documentation

data AccountSettings

A container for account-level settings in AWS Device Farm.

See: newAccountSettings smart constructor.

Constructors

AccountSettings'

Fields

awsAccountNumber :: Maybe Text

The AWS account number specified in the AccountSettings container.

maxSlots :: Maybe (HashMap Text Int)

The maximum number of device slots that the AWS account can purchase. Each maximum is expressed as an offering-id:number pair, where the offering-id represents one of the IDs returned by the ListOfferings command.

trialMinutes :: Maybe TrialMinutes

Information about an AWS account's usage of free trial device minutes.

skipAppResign :: Maybe Bool

When set to true, for private devices, Device Farm does not sign your app again. For public devices, Device Farm always signs your apps again.

For more information about how Device Farm re-signs your apps, see Do you modify my app? in the AWS Device Farm FAQs.

maxJobTimeoutMinutes :: Maybe Int

The maximum number of minutes a test run executes before it times out.

unmeteredDevices :: Maybe (HashMap DevicePlatform Int)

Returns the unmetered devices you have purchased or want to purchase.

unmeteredRemoteAccessDevices :: Maybe (HashMap DevicePlatform Int)

Returns the unmetered remote access devices you have purchased or want to purchase.

defaultJobTimeoutMinutes :: Maybe Int

The default number of minutes (at the account level) a test run executes before it times out. The default value is 150 minutes.

newAccountSettings :: AccountSettings

Create a value of AccountSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:awsAccountNumber:AccountSettings', accountSettings_awsAccountNumber - The AWS account number specified in the AccountSettings container.

$sel:maxSlots:AccountSettings', accountSettings_maxSlots - The maximum number of device slots that the AWS account can purchase. Each maximum is expressed as an offering-id:number pair, where the offering-id represents one of the IDs returned by the ListOfferings command.

$sel:trialMinutes:AccountSettings', accountSettings_trialMinutes - Information about an AWS account's usage of free trial device minutes.

$sel:skipAppResign:AccountSettings', accountSettings_skipAppResign - When set to true, for private devices, Device Farm does not sign your app again. For public devices, Device Farm always signs your apps again.

For more information about how Device Farm re-signs your apps, see Do you modify my app? in the AWS Device Farm FAQs.

$sel:maxJobTimeoutMinutes:AccountSettings', accountSettings_maxJobTimeoutMinutes - The maximum number of minutes a test run executes before it times out.

$sel:unmeteredDevices:AccountSettings', accountSettings_unmeteredDevices - Returns the unmetered devices you have purchased or want to purchase.

$sel:unmeteredRemoteAccessDevices:AccountSettings', accountSettings_unmeteredRemoteAccessDevices - Returns the unmetered remote access devices you have purchased or want to purchase.

$sel:defaultJobTimeoutMinutes:AccountSettings', accountSettings_defaultJobTimeoutMinutes - The default number of minutes (at the account level) a test run executes before it times out. The default value is 150 minutes.

accountSettings_awsAccountNumber :: Lens' AccountSettings (Maybe Text)

The AWS account number specified in the AccountSettings container.

accountSettings_maxSlots :: Lens' AccountSettings (Maybe (HashMap Text Int))

The maximum number of device slots that the AWS account can purchase. Each maximum is expressed as an offering-id:number pair, where the offering-id represents one of the IDs returned by the ListOfferings command.

accountSettings_trialMinutes :: Lens' AccountSettings (Maybe TrialMinutes)

Information about an AWS account's usage of free trial device minutes.

accountSettings_skipAppResign :: Lens' AccountSettings (Maybe Bool)

When set to true, for private devices, Device Farm does not sign your app again. For public devices, Device Farm always signs your apps again.

For more information about how Device Farm re-signs your apps, see Do you modify my app? in the AWS Device Farm FAQs.

accountSettings_maxJobTimeoutMinutes :: Lens' AccountSettings (Maybe Int)

The maximum number of minutes a test run executes before it times out.

accountSettings_unmeteredDevices :: Lens' AccountSettings (Maybe (HashMap DevicePlatform Int))

Returns the unmetered devices you have purchased or want to purchase.

accountSettings_unmeteredRemoteAccessDevices :: Lens' AccountSettings (Maybe (HashMap DevicePlatform Int))

Returns the unmetered remote access devices you have purchased or want to purchase.

accountSettings_defaultJobTimeoutMinutes :: Lens' AccountSettings (Maybe Int)

The default number of minutes (at the account level) a test run executes before it times out. The default value is 150 minutes.