Network.AWS.DeviceFarm.Types.Counters

Description

Documentation

data Counters

Represents entity counters.

See: newCounters smart constructor.

Constructors

Counters'

Fields

passed :: Maybe Int

The number of passed entities.

skipped :: Maybe Int

The number of skipped entities.

warned :: Maybe Int

The number of warned entities.

stopped :: Maybe Int

The number of stopped entities.

total :: Maybe Int

The total number of entities.

failed :: Maybe Int

The number of failed entities.

errored :: Maybe Int

The number of errored entities.

newCounters :: Counters

Create a value of Counters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:passed:Counters', counters_passed - The number of passed entities.

$sel:skipped:Counters', counters_skipped - The number of skipped entities.

$sel:warned:Counters', counters_warned - The number of warned entities.

$sel:stopped:Counters', counters_stopped - The number of stopped entities.

$sel:total:Counters', counters_total - The total number of entities.

$sel:failed:Counters', counters_failed - The number of failed entities.

$sel:errored:Counters', counters_errored - The number of errored entities.

counters_passed :: Lens' Counters (Maybe Int)

The number of passed entities.

counters_skipped :: Lens' Counters (Maybe Int)

The number of skipped entities.

counters_warned :: Lens' Counters (Maybe Int)

The number of warned entities.

counters_stopped :: Lens' Counters (Maybe Int)

The number of stopped entities.

counters_total :: Lens' Counters (Maybe Int)

The total number of entities.

counters_failed :: Lens' Counters (Maybe Int)

The number of failed entities.

counters_errored :: Lens' Counters (Maybe Int)

The number of errored entities.