Network.AWS.DeviceFarm.Types.CustomerArtifactPaths

Description

Documentation

data CustomerArtifactPaths

A JSON object that specifies the paths where the artifacts generated by the customer's tests, on the device or in the test environment, are pulled from.

Specify deviceHostPaths and optionally specify either iosPaths or androidPaths.

For web app tests, you can specify both iosPaths and androidPaths.

See: newCustomerArtifactPaths smart constructor.

Constructors

CustomerArtifactPaths'

Fields

deviceHostPaths :: Maybe [Text]

Comma-separated list of paths in the test execution environment where the artifacts generated by the customer's tests are pulled from.

iosPaths :: Maybe [Text]

Comma-separated list of paths on the iOS device where the artifacts generated by the customer's tests are pulled from.

androidPaths :: Maybe [Text]

Comma-separated list of paths on the Android device where the artifacts generated by the customer's tests are pulled from.

newCustomerArtifactPaths :: CustomerArtifactPaths

Create a value of CustomerArtifactPaths with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceHostPaths:CustomerArtifactPaths', customerArtifactPaths_deviceHostPaths - Comma-separated list of paths in the test execution environment where the artifacts generated by the customer's tests are pulled from.

$sel:iosPaths:CustomerArtifactPaths', customerArtifactPaths_iosPaths - Comma-separated list of paths on the iOS device where the artifacts generated by the customer's tests are pulled from.

$sel:androidPaths:CustomerArtifactPaths', customerArtifactPaths_androidPaths - Comma-separated list of paths on the Android device where the artifacts generated by the customer's tests are pulled from.

customerArtifactPaths_deviceHostPaths :: Lens' CustomerArtifactPaths (Maybe [Text])

Comma-separated list of paths in the test execution environment where the artifacts generated by the customer's tests are pulled from.

customerArtifactPaths_iosPaths :: Lens' CustomerArtifactPaths (Maybe [Text])

Comma-separated list of paths on the iOS device where the artifacts generated by the customer's tests are pulled from.

customerArtifactPaths_androidPaths :: Lens' CustomerArtifactPaths (Maybe [Text])

Comma-separated list of paths on the Android device where the artifacts generated by the customer's tests are pulled from.