Network.AWS.DeviceFarm.Types.Device

Description

Documentation

data Device

Represents a device type that an app is tested against.

See: newDevice smart constructor.

Constructors

Device'

Fields

platform :: Maybe DevicePlatform

The device's platform.

Allowed values include:

ANDROID

IOS

manufacturer :: Maybe Text

The device's manufacturer name.

model :: Maybe Text

The device's model name.

memory :: Maybe Integer

The device's total memory size, expressed in bytes.

availability :: Maybe DeviceAvailability

Indicates how likely a device is available for a test run. Currently available in the ListDevices and GetDevice API methods.

fleetName :: Maybe Text

The name of the fleet to which this device belongs.

fleetType :: Maybe Text

The type of fleet to which this device belongs. Possible values are PRIVATE and PUBLIC.

formFactor :: Maybe DeviceFormFactor

The device's form factor.

Allowed values include:

PHONE

TABLET

arn :: Maybe Text

The device's ARN.

remoteAccessEnabled :: Maybe Bool

Specifies whether remote access has been enabled for the specified device.

instances :: Maybe [DeviceInstance]

The instances that belong to this device.

name :: Maybe Text

The device's display name.

image :: Maybe Text

The device's image name.

carrier :: Maybe Text

The device's carrier.

os :: Maybe Text

The device's operating system type.

heapSize :: Maybe Integer

The device's heap size, expressed in bytes.

radio :: Maybe Text

The device's radio.

resolution :: Maybe Resolution

The resolution of the device.

cpu :: Maybe CPU

Information about the device's CPU.

remoteDebugEnabled :: Maybe Bool

This flag is set to true if remote debugging is enabled for the device.

Remote debugging is no longer supported.

modelId :: Maybe Text

The device's model ID.

newDevice :: Device

Create a value of Device with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:platform:Device', device_platform - The device's platform.

Allowed values include:

ANDROID

IOS

$sel:manufacturer:Device', device_manufacturer - The device's manufacturer name.

$sel:model:Device', device_model - The device's model name.

$sel:memory:Device', device_memory - The device's total memory size, expressed in bytes.

$sel:availability:Device', device_availability - Indicates how likely a device is available for a test run. Currently available in the ListDevices and GetDevice API methods.

$sel:fleetName:Device', device_fleetName - The name of the fleet to which this device belongs.

$sel:fleetType:Device', device_fleetType - The type of fleet to which this device belongs. Possible values are PRIVATE and PUBLIC.

$sel:formFactor:Device', device_formFactor - The device's form factor.

Allowed values include:

PHONE

TABLET

$sel:arn:Device', device_arn - The device's ARN.

$sel:remoteAccessEnabled:Device', device_remoteAccessEnabled - Specifies whether remote access has been enabled for the specified device.

$sel:instances:Device', device_instances - The instances that belong to this device.

$sel:name:Device', device_name - The device's display name.

$sel:image:Device', device_image - The device's image name.

$sel:carrier:Device', device_carrier - The device's carrier.

$sel:os:Device', device_os - The device's operating system type.

$sel:heapSize:Device', device_heapSize - The device's heap size, expressed in bytes.

$sel:radio:Device', device_radio - The device's radio.

$sel:resolution:Device', device_resolution - The resolution of the device.

$sel:cpu:Device', device_cpu - Information about the device's CPU.

$sel:remoteDebugEnabled:Device', device_remoteDebugEnabled - This flag is set to true if remote debugging is enabled for the device.

Remote debugging is no longer supported.

$sel:modelId:Device', device_modelId - The device's model ID.

device_platform :: Lens' Device (Maybe DevicePlatform)

The device's platform.

Allowed values include:

ANDROID

IOS

device_manufacturer :: Lens' Device (Maybe Text)

The device's manufacturer name.

device_model :: Lens' Device (Maybe Text)

The device's model name.

device_memory :: Lens' Device (Maybe Integer)

The device's total memory size, expressed in bytes.

device_availability :: Lens' Device (Maybe DeviceAvailability)

Indicates how likely a device is available for a test run. Currently available in the ListDevices and GetDevice API methods.

device_fleetName :: Lens' Device (Maybe Text)

The name of the fleet to which this device belongs.

device_fleetType :: Lens' Device (Maybe Text)

The type of fleet to which this device belongs. Possible values are PRIVATE and PUBLIC.

device_formFactor :: Lens' Device (Maybe DeviceFormFactor)

The device's form factor.

Allowed values include:

PHONE

TABLET

device_arn :: Lens' Device (Maybe Text)

The device's ARN.

device_remoteAccessEnabled :: Lens' Device (Maybe Bool)

Specifies whether remote access has been enabled for the specified device.

device_instances :: Lens' Device (Maybe [DeviceInstance])

The instances that belong to this device.

device_name :: Lens' Device (Maybe Text)

The device's display name.

device_image :: Lens' Device (Maybe Text)

The device's image name.

device_carrier :: Lens' Device (Maybe Text)

The device's carrier.

device_os :: Lens' Device (Maybe Text)

The device's operating system type.

device_heapSize :: Lens' Device (Maybe Integer)

The device's heap size, expressed in bytes.

device_radio :: Lens' Device (Maybe Text)

The device's radio.

device_resolution :: Lens' Device (Maybe Resolution)

The resolution of the device.

device_cpu :: Lens' Device (Maybe CPU)

Information about the device's CPU.

device_remoteDebugEnabled :: Lens' Device (Maybe Bool)

This flag is set to true if remote debugging is enabled for the device.

Remote debugging is no longer supported.

device_modelId :: Lens' Device (Maybe Text)

The device's model ID.