Network.AWS.DeviceFarm.Types.ExecutionConfiguration

Description

Documentation

data ExecutionConfiguration

Represents configuration information about a test run, such as the execution timeout (in minutes).

See: newExecutionConfiguration smart constructor.

Constructors

ExecutionConfiguration'

Fields

appPackagesCleanup :: Maybe Bool

True if app package cleanup is enabled at the beginning of the test. Otherwise, false.

videoCapture :: Maybe Bool

Set to true to enable video capture. Otherwise, set to false. The default is true.

skipAppResign :: Maybe Bool

When set to true, for private devices, Device Farm does not sign your app again. For public devices, Device Farm always signs your apps again.

For more information about how Device Farm re-signs your apps, see Do you modify my app? in the AWS Device Farm FAQs.

jobTimeoutMinutes :: Maybe Int

The number of minutes a test run executes before it times out.

accountsCleanup :: Maybe Bool

True if account cleanup is enabled at the beginning of the test. Otherwise, false.

newExecutionConfiguration :: ExecutionConfiguration

Create a value of ExecutionConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appPackagesCleanup:ExecutionConfiguration', executionConfiguration_appPackagesCleanup - True if app package cleanup is enabled at the beginning of the test. Otherwise, false.

$sel:videoCapture:ExecutionConfiguration', executionConfiguration_videoCapture - Set to true to enable video capture. Otherwise, set to false. The default is true.

$sel:skipAppResign:ExecutionConfiguration', executionConfiguration_skipAppResign - When set to true, for private devices, Device Farm does not sign your app again. For public devices, Device Farm always signs your apps again.

For more information about how Device Farm re-signs your apps, see Do you modify my app? in the AWS Device Farm FAQs.

$sel:jobTimeoutMinutes:ExecutionConfiguration', executionConfiguration_jobTimeoutMinutes - The number of minutes a test run executes before it times out.

$sel:accountsCleanup:ExecutionConfiguration', executionConfiguration_accountsCleanup - True if account cleanup is enabled at the beginning of the test. Otherwise, false.

executionConfiguration_appPackagesCleanup :: Lens' ExecutionConfiguration (Maybe Bool)

True if app package cleanup is enabled at the beginning of the test. Otherwise, false.

executionConfiguration_videoCapture :: Lens' ExecutionConfiguration (Maybe Bool)

Set to true to enable video capture. Otherwise, set to false. The default is true.

executionConfiguration_skipAppResign :: Lens' ExecutionConfiguration (Maybe Bool)

When set to true, for private devices, Device Farm does not sign your app again. For public devices, Device Farm always signs your apps again.

For more information about how Device Farm re-signs your apps, see Do you modify my app? in the AWS Device Farm FAQs.

executionConfiguration_jobTimeoutMinutes :: Lens' ExecutionConfiguration (Maybe Int)

The number of minutes a test run executes before it times out.

executionConfiguration_accountsCleanup :: Lens' ExecutionConfiguration (Maybe Bool)

True if account cleanup is enabled at the beginning of the test. Otherwise, false.