Network.AWS.DeviceFarm.Types.InstanceProfile

Description

Documentation

data InstanceProfile

Represents the instance profile.

See: newInstanceProfile smart constructor.

Constructors

InstanceProfile'

Fields

excludeAppPackagesFromCleanup :: Maybe [Text]

An array of strings containing the list of app packages that should not be cleaned up from the device after a test run completes.

The list of packages is considered only if you set packageCleanup to true.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the instance profile.

name :: Maybe Text

The name of the instance profile.

description :: Maybe Text

The description of the instance profile.

rebootAfterUse :: Maybe Bool

When set to true, Device Farm reboots the instance after a test run. The default value is true.

packageCleanup :: Maybe Bool

When set to true, Device Farm removes app packages after a test run. The default value is false for private devices.

newInstanceProfile :: InstanceProfile

Create a value of InstanceProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:excludeAppPackagesFromCleanup:InstanceProfile', instanceProfile_excludeAppPackagesFromCleanup - An array of strings containing the list of app packages that should not be cleaned up from the device after a test run completes.

The list of packages is considered only if you set packageCleanup to true.

$sel:arn:InstanceProfile', instanceProfile_arn - The Amazon Resource Name (ARN) of the instance profile.

$sel:name:InstanceProfile', instanceProfile_name - The name of the instance profile.

$sel:description:InstanceProfile', instanceProfile_description - The description of the instance profile.

$sel:rebootAfterUse:InstanceProfile', instanceProfile_rebootAfterUse - When set to true, Device Farm reboots the instance after a test run. The default value is true.

$sel:packageCleanup:InstanceProfile', instanceProfile_packageCleanup - When set to true, Device Farm removes app packages after a test run. The default value is false for private devices.

instanceProfile_excludeAppPackagesFromCleanup :: Lens' InstanceProfile (Maybe [Text])

An array of strings containing the list of app packages that should not be cleaned up from the device after a test run completes.

The list of packages is considered only if you set packageCleanup to true.

instanceProfile_arn :: Lens' InstanceProfile (Maybe Text)

The Amazon Resource Name (ARN) of the instance profile.

instanceProfile_name :: Lens' InstanceProfile (Maybe Text)

The name of the instance profile.

instanceProfile_description :: Lens' InstanceProfile (Maybe Text)

The description of the instance profile.

instanceProfile_rebootAfterUse :: Lens' InstanceProfile (Maybe Bool)

When set to true, Device Farm reboots the instance after a test run. The default value is true.

instanceProfile_packageCleanup :: Lens' InstanceProfile (Maybe Bool)

When set to true, Device Farm removes app packages after a test run. The default value is false for private devices.