Network.AWS.DeviceFarm.Types.Job

Description

Documentation

data Job

Represents a device.

See: newJob smart constructor.

Constructors

Job'

Fields

instanceArn :: Maybe Text

The ARN of the instance.

status :: Maybe ExecutionStatus

The job's status.

Allowed values include:

PENDING

PENDING_CONCURRENCY

PENDING_DEVICE

PROCESSING

SCHEDULING

PREPARING

RUNNING

COMPLETED

STOPPING

counters :: Maybe Counters

The job's result counters.

arn :: Maybe Text

The job's ARN.

created :: Maybe POSIX

When the job was created.

device :: Maybe Device

The device (phone or tablet).

stopped :: Maybe POSIX

The job's stop time.

result :: Maybe ExecutionResult

The job's result.

Allowed values include:

PENDING

PASSED

WARNED

FAILED

SKIPPED

ERRORED

STOPPED

name :: Maybe Text

The job's name.

videoEndpoint :: Maybe Text

The endpoint for streaming device video.

deviceMinutes :: Maybe DeviceMinutes

Represents the total (metered or unmetered) minutes used by the job.

videoCapture :: Maybe Bool

This value is set to true if video capture is enabled. Otherwise, it is set to false.

type' :: Maybe TestType

The job's type.

Allowed values include the following:

BUILTIN_FUZZ

BUILTIN_EXPLORER. For Android, an app explorer that traverses an Android app, interacting with it and capturing screenshots at the same time.

APPIUM_JAVA_JUNIT

APPIUM_JAVA_TESTNG

APPIUM_PYTHON

APPIUM_NODE

APPIUM_RUBY

APPIUM_WEB_JAVA_JUNIT

APPIUM_WEB_JAVA_TESTNG

APPIUM_WEB_PYTHON

APPIUM_WEB_NODE

APPIUM_WEB_RUBY

CALABASH

INSTRUMENTATION

UIAUTOMATION

UIAUTOMATOR

XCTEST

XCTEST_UI

message :: Maybe Text

A message about the job's result.

started :: Maybe POSIX

The job's start time.

newJob :: Job

Create a value of Job with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceArn:Job', job_instanceArn - The ARN of the instance.

$sel:status:Job', job_status - The job's status.

Allowed values include:

PENDING

PENDING_CONCURRENCY

PENDING_DEVICE

PROCESSING

SCHEDULING

PREPARING

RUNNING

COMPLETED

STOPPING

$sel:counters:Job', job_counters - The job's result counters.

$sel:arn:Job', job_arn - The job's ARN.

$sel:created:Job', job_created - When the job was created.

$sel:device:Job', job_device - The device (phone or tablet).

$sel:stopped:Job', job_stopped - The job's stop time.

$sel:result:Job', job_result - The job's result.

Allowed values include:

PENDING

PASSED

WARNED

FAILED

SKIPPED

ERRORED

STOPPED

$sel:name:Job', job_name - The job's name.

$sel:videoEndpoint:Job', job_videoEndpoint - The endpoint for streaming device video.

$sel:deviceMinutes:Job', job_deviceMinutes - Represents the total (metered or unmetered) minutes used by the job.

$sel:videoCapture:Job', job_videoCapture - This value is set to true if video capture is enabled. Otherwise, it is set to false.

$sel:type':Job', job_type - The job's type.

Allowed values include the following:

BUILTIN_FUZZ

BUILTIN_EXPLORER. For Android, an app explorer that traverses an Android app, interacting with it and capturing screenshots at the same time.

APPIUM_JAVA_JUNIT

APPIUM_JAVA_TESTNG

APPIUM_PYTHON

APPIUM_NODE

APPIUM_RUBY

APPIUM_WEB_JAVA_JUNIT

APPIUM_WEB_JAVA_TESTNG

APPIUM_WEB_PYTHON

APPIUM_WEB_NODE

APPIUM_WEB_RUBY

CALABASH

INSTRUMENTATION

UIAUTOMATION

UIAUTOMATOR

XCTEST

XCTEST_UI

$sel:message:Job', job_message - A message about the job's result.

$sel:started:Job', job_started - The job's start time.

job_instanceArn :: Lens' Job (Maybe Text)

The ARN of the instance.

job_status :: Lens' Job (Maybe ExecutionStatus)

The job's status.

Allowed values include:

PENDING

PENDING_CONCURRENCY

PENDING_DEVICE

PROCESSING

SCHEDULING

PREPARING

RUNNING

COMPLETED

STOPPING

job_counters :: Lens' Job (Maybe Counters)

The job's result counters.

job_arn :: Lens' Job (Maybe Text)

The job's ARN.

job_created :: Lens' Job (Maybe UTCTime)

When the job was created.

job_device :: Lens' Job (Maybe Device)

The device (phone or tablet).

job_stopped :: Lens' Job (Maybe UTCTime)

The job's stop time.

job_result :: Lens' Job (Maybe ExecutionResult)

The job's result.

Allowed values include:

PENDING

PASSED

WARNED

FAILED

SKIPPED

ERRORED

STOPPED

job_name :: Lens' Job (Maybe Text)

The job's name.

job_videoEndpoint :: Lens' Job (Maybe Text)

The endpoint for streaming device video.

job_deviceMinutes :: Lens' Job (Maybe DeviceMinutes)

Represents the total (metered or unmetered) minutes used by the job.

job_videoCapture :: Lens' Job (Maybe Bool)

This value is set to true if video capture is enabled. Otherwise, it is set to false.

job_type :: Lens' Job (Maybe TestType)

The job's type.

Allowed values include the following:

BUILTIN_FUZZ

BUILTIN_EXPLORER. For Android, an app explorer that traverses an Android app, interacting with it and capturing screenshots at the same time.

APPIUM_JAVA_JUNIT

APPIUM_JAVA_TESTNG

APPIUM_PYTHON

APPIUM_NODE

APPIUM_RUBY

APPIUM_WEB_JAVA_JUNIT

APPIUM_WEB_JAVA_TESTNG

APPIUM_WEB_PYTHON

APPIUM_WEB_NODE

APPIUM_WEB_RUBY

CALABASH

INSTRUMENTATION

UIAUTOMATION

UIAUTOMATOR

XCTEST

XCTEST_UI

job_message :: Lens' Job (Maybe Text)

A message about the job's result.

job_started :: Lens' Job (Maybe UTCTime)

The job's start time.