Network.AWS.DeviceFarm.Types.MonetaryAmount

Description

Documentation

data MonetaryAmount

A number that represents the monetary amount for an offering or transaction.

See: newMonetaryAmount smart constructor.

Constructors

MonetaryAmount'

Fields

amount :: Maybe Double

The numerical amount of an offering or transaction.

currencyCode :: Maybe CurrencyCode

The currency code of a monetary amount. For example, USD means U.S. dollars.

newMonetaryAmount :: MonetaryAmount

Create a value of MonetaryAmount with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:amount:MonetaryAmount', monetaryAmount_amount - The numerical amount of an offering or transaction.

$sel:currencyCode:MonetaryAmount', monetaryAmount_currencyCode - The currency code of a monetary amount. For example, USD means U.S. dollars.

monetaryAmount_amount :: Lens' MonetaryAmount (Maybe Double)

The numerical amount of an offering or transaction.

monetaryAmount_currencyCode :: Lens' MonetaryAmount (Maybe CurrencyCode)

The currency code of a monetary amount. For example, USD means U.S. dollars.