Network.AWS.DeviceFarm.Types.NetworkProfile

Description

Documentation

data NetworkProfile

An array of settings that describes characteristics of a network profile.

See: newNetworkProfile smart constructor.

Constructors

NetworkProfile'

Fields

uplinkJitterMs :: Maybe Integer

Time variation in the delay of received packets in milliseconds as an integer from 0 to 2000.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the network profile.

uplinkLossPercent :: Maybe Natural

Proportion of transmitted packets that fail to arrive from 0 to 100 percent.

downlinkJitterMs :: Maybe Integer

Time variation in the delay of received packets in milliseconds as an integer from 0 to 2000.

name :: Maybe Text

The name of the network profile.

downlinkLossPercent :: Maybe Natural

Proportion of received packets that fail to arrive from 0 to 100 percent.

type' :: Maybe NetworkProfileType

The type of network profile. Valid values are listed here.

uplinkDelayMs :: Maybe Integer

Delay time for all packets to destination in milliseconds as an integer from 0 to 2000.

uplinkBandwidthBits :: Maybe Integer

The data throughput rate in bits per second, as an integer from 0 to 104857600.

description :: Maybe Text

The description of the network profile.

downlinkDelayMs :: Maybe Integer

Delay time for all packets to destination in milliseconds as an integer from 0 to 2000.

downlinkBandwidthBits :: Maybe Integer

The data throughput rate in bits per second, as an integer from 0 to 104857600.

newNetworkProfile :: NetworkProfile

Create a value of NetworkProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:uplinkJitterMs:NetworkProfile', networkProfile_uplinkJitterMs - Time variation in the delay of received packets in milliseconds as an integer from 0 to 2000.

$sel:arn:NetworkProfile', networkProfile_arn - The Amazon Resource Name (ARN) of the network profile.

$sel:uplinkLossPercent:NetworkProfile', networkProfile_uplinkLossPercent - Proportion of transmitted packets that fail to arrive from 0 to 100 percent.

$sel:downlinkJitterMs:NetworkProfile', networkProfile_downlinkJitterMs - Time variation in the delay of received packets in milliseconds as an integer from 0 to 2000.

$sel:name:NetworkProfile', networkProfile_name - The name of the network profile.

$sel:downlinkLossPercent:NetworkProfile', networkProfile_downlinkLossPercent - Proportion of received packets that fail to arrive from 0 to 100 percent.

$sel:type':NetworkProfile', networkProfile_type - The type of network profile. Valid values are listed here.

$sel:uplinkDelayMs:NetworkProfile', networkProfile_uplinkDelayMs - Delay time for all packets to destination in milliseconds as an integer from 0 to 2000.

$sel:uplinkBandwidthBits:NetworkProfile', networkProfile_uplinkBandwidthBits - The data throughput rate in bits per second, as an integer from 0 to 104857600.

$sel:description:NetworkProfile', networkProfile_description - The description of the network profile.

$sel:downlinkDelayMs:NetworkProfile', networkProfile_downlinkDelayMs - Delay time for all packets to destination in milliseconds as an integer from 0 to 2000.

$sel:downlinkBandwidthBits:NetworkProfile', networkProfile_downlinkBandwidthBits - The data throughput rate in bits per second, as an integer from 0 to 104857600.

networkProfile_uplinkJitterMs :: Lens' NetworkProfile (Maybe Integer)

Time variation in the delay of received packets in milliseconds as an integer from 0 to 2000.

networkProfile_arn :: Lens' NetworkProfile (Maybe Text)

The Amazon Resource Name (ARN) of the network profile.

networkProfile_uplinkLossPercent :: Lens' NetworkProfile (Maybe Natural)

Proportion of transmitted packets that fail to arrive from 0 to 100 percent.

networkProfile_downlinkJitterMs :: Lens' NetworkProfile (Maybe Integer)

Time variation in the delay of received packets in milliseconds as an integer from 0 to 2000.

networkProfile_name :: Lens' NetworkProfile (Maybe Text)

The name of the network profile.

networkProfile_downlinkLossPercent :: Lens' NetworkProfile (Maybe Natural)

Proportion of received packets that fail to arrive from 0 to 100 percent.

networkProfile_type :: Lens' NetworkProfile (Maybe NetworkProfileType)

The type of network profile. Valid values are listed here.

networkProfile_uplinkDelayMs :: Lens' NetworkProfile (Maybe Integer)

Delay time for all packets to destination in milliseconds as an integer from 0 to 2000.

networkProfile_uplinkBandwidthBits :: Lens' NetworkProfile (Maybe Integer)

The data throughput rate in bits per second, as an integer from 0 to 104857600.

networkProfile_description :: Lens' NetworkProfile (Maybe Text)

The description of the network profile.

networkProfile_downlinkDelayMs :: Lens' NetworkProfile (Maybe Integer)

Delay time for all packets to destination in milliseconds as an integer from 0 to 2000.

networkProfile_downlinkBandwidthBits :: Lens' NetworkProfile (Maybe Integer)

The data throughput rate in bits per second, as an integer from 0 to 104857600.