Network.AWS.DeviceFarm.Types.Offering

Description

Documentation

data Offering

Represents the metadata of a device offering.

See: newOffering smart constructor.

Constructors

Offering'

Fields

platform :: Maybe DevicePlatform

The platform of the device (for example, ANDROID or IOS).

id :: Maybe Text

The ID that corresponds to a device offering.

description :: Maybe Text

A string that describes the offering.

type' :: Maybe OfferingType

The type of offering (for example, RECURRING) for a device.

recurringCharges :: Maybe [RecurringCharge]

Specifies whether there are recurring charges for the offering.

newOffering :: Offering

Create a value of Offering with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:platform:Offering', offering_platform - The platform of the device (for example, ANDROID or IOS).

$sel:id:Offering', offering_id - The ID that corresponds to a device offering.

$sel:description:Offering', offering_description - A string that describes the offering.

$sel:type':Offering', offering_type - The type of offering (for example, RECURRING) for a device.

$sel:recurringCharges:Offering', offering_recurringCharges - Specifies whether there are recurring charges for the offering.

offering_platform :: Lens' Offering (Maybe DevicePlatform)

The platform of the device (for example, ANDROID or IOS).

offering_id :: Lens' Offering (Maybe Text)

The ID that corresponds to a device offering.

offering_description :: Lens' Offering (Maybe Text)

A string that describes the offering.

offering_type :: Lens' Offering (Maybe OfferingType)

The type of offering (for example, RECURRING) for a device.

offering_recurringCharges :: Lens' Offering (Maybe [RecurringCharge])

Specifies whether there are recurring charges for the offering.