Network.AWS.DeviceFarm.Types.Problem

Description

Documentation

data Problem

Represents a specific warning or failure.

See: newProblem smart constructor.

Constructors

Problem'

Fields

job :: Maybe ProblemDetail

Information about the associated job.

result :: Maybe ExecutionResult

The problem's result.

Allowed values include:

PENDING

PASSED

WARNED

FAILED

SKIPPED

ERRORED

STOPPED

message :: Maybe Text

A message about the problem's result.

device :: Maybe Device

Information about the associated device.

run :: Maybe ProblemDetail

Information about the associated run.

suite :: Maybe ProblemDetail

Information about the associated suite.

test :: Maybe ProblemDetail

Information about the associated test.

newProblem :: Problem

Create a value of Problem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:job:Problem', problem_job - Information about the associated job.

$sel:result:Problem', problem_result - The problem's result.

Allowed values include:

PENDING

PASSED

WARNED

FAILED

SKIPPED

ERRORED

STOPPED

$sel:message:Problem', problem_message - A message about the problem's result.

$sel:device:Problem', problem_device - Information about the associated device.

$sel:run:Problem', problem_run - Information about the associated run.

$sel:suite:Problem', problem_suite - Information about the associated suite.

$sel:test:Problem', problem_test - Information about the associated test.

problem_job :: Lens' Problem (Maybe ProblemDetail)

Information about the associated job.

problem_result :: Lens' Problem (Maybe ExecutionResult)

The problem's result.

Allowed values include:

PENDING

PASSED

WARNED

FAILED

SKIPPED

ERRORED

STOPPED

problem_message :: Lens' Problem (Maybe Text)

A message about the problem's result.

problem_device :: Lens' Problem (Maybe Device)

Information about the associated device.

problem_run :: Lens' Problem (Maybe ProblemDetail)

Information about the associated run.

problem_suite :: Lens' Problem (Maybe ProblemDetail)

Information about the associated suite.

problem_test :: Lens' Problem (Maybe ProblemDetail)

Information about the associated test.