Network.AWS.DeviceFarm.Types.RemoteAccessSession

Description

Documentation

data RemoteAccessSession

Represents information about the remote access session.

See: newRemoteAccessSession smart constructor.

Constructors

RemoteAccessSession'

Fields

billingMethod :: Maybe BillingMethod

The billing method of the remote access session. Possible values include METERED or UNMETERED. For more information about metered devices, see AWS Device Farm terminology.

clientId :: Maybe Text

Unique identifier of your client for the remote access session. Only returned if remote debugging is enabled for the remote access session.

Remote debugging is no longer supported.

deviceUdid :: Maybe Text

Unique device identifier for the remote device. Only returned if remote debugging is enabled for the remote access session.

Remote debugging is no longer supported.

skipAppResign :: Maybe Bool

When set to true, for private devices, Device Farm does not sign your app again. For public devices, Device Farm always signs your apps again.

For more information about how Device Farm re-signs your apps, see Do you modify my app? in the AWS Device Farm FAQs.

instanceArn :: Maybe Text

The ARN of the instance.

status :: Maybe ExecutionStatus

The status of the remote access session. Can be any of the following:

PENDING.

PENDING_CONCURRENCY.

PENDING_DEVICE.

PROCESSING.

SCHEDULING.

PREPARING.

RUNNING.

COMPLETED.

STOPPING.

remoteRecordEnabled :: Maybe Bool

This flag is set to true if remote recording is enabled for the remote access session.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the remote access session.

remoteRecordAppArn :: Maybe Text

The ARN for the app to be recorded in the remote access session.

created :: Maybe POSIX

The date and time the remote access session was created.

device :: Maybe Device

The device (phone or tablet) used in the remote access session.

stopped :: Maybe POSIX

The date and time the remote access session was stopped.

result :: Maybe ExecutionResult

The result of the remote access session. Can be any of the following:

PENDING.

PASSED.

WARNED.

FAILED.

SKIPPED.

ERRORED.

STOPPED.

name :: Maybe Text

The name of the remote access session.

deviceMinutes :: Maybe DeviceMinutes

The number of minutes a device is used in a remote access session (including setup and teardown minutes).

remoteDebugEnabled :: Maybe Bool

This flag is set to true if remote debugging is enabled for the remote access session.

Remote debugging is no longer supported.

endpoint :: Maybe Text

The endpoint for the remote access sesssion.

message :: Maybe Text

A message about the remote access session.

hostAddress :: Maybe Text

IP address of the EC2 host where you need to connect to remotely debug devices. Only returned if remote debugging is enabled for the remote access session.

Remote debugging is no longer supported.

interactionMode :: Maybe InteractionMode

The interaction mode of the remote access session. Valid values are:

INTERACTIVE: You can interact with the iOS device by viewing, touching, and rotating the screen. You cannot run XCUITest framework-based tests in this mode.

NO_VIDEO: You are connected to the device, but cannot interact with it or view the screen. This mode has the fastest test execution speed. You can run XCUITest framework-based tests in this mode.

VIDEO_ONLY: You can view the screen, but cannot touch or rotate it. You can run XCUITest framework-based tests and watch the screen in this mode.

started :: Maybe POSIX

The date and time the remote access session was started.

newRemoteAccessSession :: RemoteAccessSession

Create a value of RemoteAccessSession with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:billingMethod:RemoteAccessSession', remoteAccessSession_billingMethod - The billing method of the remote access session. Possible values include METERED or UNMETERED. For more information about metered devices, see AWS Device Farm terminology.

$sel:clientId:RemoteAccessSession', remoteAccessSession_clientId - Unique identifier of your client for the remote access session. Only returned if remote debugging is enabled for the remote access session.

Remote debugging is no longer supported.

$sel:deviceUdid:RemoteAccessSession', remoteAccessSession_deviceUdid - Unique device identifier for the remote device. Only returned if remote debugging is enabled for the remote access session.

Remote debugging is no longer supported.

$sel:skipAppResign:RemoteAccessSession', remoteAccessSession_skipAppResign - When set to true, for private devices, Device Farm does not sign your app again. For public devices, Device Farm always signs your apps again.

For more information about how Device Farm re-signs your apps, see Do you modify my app? in the AWS Device Farm FAQs.

$sel:instanceArn:RemoteAccessSession', remoteAccessSession_instanceArn - The ARN of the instance.

$sel:status:RemoteAccessSession', remoteAccessSession_status - The status of the remote access session. Can be any of the following:

PENDING.

PENDING_CONCURRENCY.

PENDING_DEVICE.

PROCESSING.

SCHEDULING.

PREPARING.

RUNNING.

COMPLETED.

STOPPING.

$sel:remoteRecordEnabled:RemoteAccessSession', remoteAccessSession_remoteRecordEnabled - This flag is set to true if remote recording is enabled for the remote access session.

$sel:arn:RemoteAccessSession', remoteAccessSession_arn - The Amazon Resource Name (ARN) of the remote access session.

$sel:remoteRecordAppArn:RemoteAccessSession', remoteAccessSession_remoteRecordAppArn - The ARN for the app to be recorded in the remote access session.

$sel:created:RemoteAccessSession', remoteAccessSession_created - The date and time the remote access session was created.

$sel:device:RemoteAccessSession', remoteAccessSession_device - The device (phone or tablet) used in the remote access session.

$sel:stopped:RemoteAccessSession', remoteAccessSession_stopped - The date and time the remote access session was stopped.

$sel:result:RemoteAccessSession', remoteAccessSession_result - The result of the remote access session. Can be any of the following:

PENDING.

PASSED.

WARNED.

FAILED.

SKIPPED.

ERRORED.

STOPPED.

$sel:name:RemoteAccessSession', remoteAccessSession_name - The name of the remote access session.

$sel:deviceMinutes:RemoteAccessSession', remoteAccessSession_deviceMinutes - The number of minutes a device is used in a remote access session (including setup and teardown minutes).

$sel:remoteDebugEnabled:RemoteAccessSession', remoteAccessSession_remoteDebugEnabled - This flag is set to true if remote debugging is enabled for the remote access session.

Remote debugging is no longer supported.

$sel:endpoint:RemoteAccessSession', remoteAccessSession_endpoint - The endpoint for the remote access sesssion.

$sel:message:RemoteAccessSession', remoteAccessSession_message - A message about the remote access session.

$sel:hostAddress:RemoteAccessSession', remoteAccessSession_hostAddress - IP address of the EC2 host where you need to connect to remotely debug devices. Only returned if remote debugging is enabled for the remote access session.

Remote debugging is no longer supported.

$sel:interactionMode:RemoteAccessSession', remoteAccessSession_interactionMode - The interaction mode of the remote access session. Valid values are:

INTERACTIVE: You can interact with the iOS device by viewing, touching, and rotating the screen. You cannot run XCUITest framework-based tests in this mode.

NO_VIDEO: You are connected to the device, but cannot interact with it or view the screen. This mode has the fastest test execution speed. You can run XCUITest framework-based tests in this mode.

VIDEO_ONLY: You can view the screen, but cannot touch or rotate it. You can run XCUITest framework-based tests and watch the screen in this mode.

$sel:started:RemoteAccessSession', remoteAccessSession_started - The date and time the remote access session was started.

remoteAccessSession_billingMethod :: Lens' RemoteAccessSession (Maybe BillingMethod)

The billing method of the remote access session. Possible values include METERED or UNMETERED. For more information about metered devices, see AWS Device Farm terminology.

remoteAccessSession_clientId :: Lens' RemoteAccessSession (Maybe Text)

Unique identifier of your client for the remote access session. Only returned if remote debugging is enabled for the remote access session.

Remote debugging is no longer supported.

remoteAccessSession_deviceUdid :: Lens' RemoteAccessSession (Maybe Text)

Unique device identifier for the remote device. Only returned if remote debugging is enabled for the remote access session.

Remote debugging is no longer supported.

remoteAccessSession_skipAppResign :: Lens' RemoteAccessSession (Maybe Bool)

When set to true, for private devices, Device Farm does not sign your app again. For public devices, Device Farm always signs your apps again.

For more information about how Device Farm re-signs your apps, see Do you modify my app? in the AWS Device Farm FAQs.

remoteAccessSession_instanceArn :: Lens' RemoteAccessSession (Maybe Text)

The ARN of the instance.

remoteAccessSession_status :: Lens' RemoteAccessSession (Maybe ExecutionStatus)

The status of the remote access session. Can be any of the following:

PENDING.

PENDING_CONCURRENCY.

PENDING_DEVICE.

PROCESSING.

SCHEDULING.

PREPARING.

RUNNING.

COMPLETED.

STOPPING.

remoteAccessSession_remoteRecordEnabled :: Lens' RemoteAccessSession (Maybe Bool)

This flag is set to true if remote recording is enabled for the remote access session.

remoteAccessSession_arn :: Lens' RemoteAccessSession (Maybe Text)

The Amazon Resource Name (ARN) of the remote access session.

remoteAccessSession_remoteRecordAppArn :: Lens' RemoteAccessSession (Maybe Text)

The ARN for the app to be recorded in the remote access session.

remoteAccessSession_created :: Lens' RemoteAccessSession (Maybe UTCTime)

The date and time the remote access session was created.

remoteAccessSession_device :: Lens' RemoteAccessSession (Maybe Device)

The device (phone or tablet) used in the remote access session.

remoteAccessSession_stopped :: Lens' RemoteAccessSession (Maybe UTCTime)

The date and time the remote access session was stopped.

remoteAccessSession_result :: Lens' RemoteAccessSession (Maybe ExecutionResult)

The result of the remote access session. Can be any of the following:

PENDING.

PASSED.

WARNED.

FAILED.

SKIPPED.

ERRORED.

STOPPED.

remoteAccessSession_name :: Lens' RemoteAccessSession (Maybe Text)

The name of the remote access session.

remoteAccessSession_deviceMinutes :: Lens' RemoteAccessSession (Maybe DeviceMinutes)

The number of minutes a device is used in a remote access session (including setup and teardown minutes).

remoteAccessSession_remoteDebugEnabled :: Lens' RemoteAccessSession (Maybe Bool)

This flag is set to true if remote debugging is enabled for the remote access session.

Remote debugging is no longer supported.

remoteAccessSession_endpoint :: Lens' RemoteAccessSession (Maybe Text)

The endpoint for the remote access sesssion.

remoteAccessSession_message :: Lens' RemoteAccessSession (Maybe Text)

A message about the remote access session.

remoteAccessSession_hostAddress :: Lens' RemoteAccessSession (Maybe Text)

IP address of the EC2 host where you need to connect to remotely debug devices. Only returned if remote debugging is enabled for the remote access session.

Remote debugging is no longer supported.

remoteAccessSession_interactionMode :: Lens' RemoteAccessSession (Maybe InteractionMode)

The interaction mode of the remote access session. Valid values are:

INTERACTIVE: You can interact with the iOS device by viewing, touching, and rotating the screen. You cannot run XCUITest framework-based tests in this mode.

NO_VIDEO: You are connected to the device, but cannot interact with it or view the screen. This mode has the fastest test execution speed. You can run XCUITest framework-based tests in this mode.

VIDEO_ONLY: You can view the screen, but cannot touch or rotate it. You can run XCUITest framework-based tests and watch the screen in this mode.

remoteAccessSession_started :: Lens' RemoteAccessSession (Maybe UTCTime)

The date and time the remote access session was started.