Network.AWS.DeviceFarm.Types.Run

Description

Documentation

data Run

Represents a test run on a set of devices with a given app package, test parameters, and so on.

See: newRun smart constructor.

Constructors

Run'

Fields

billingMethod :: Maybe BillingMethod

Specifies the billing method for a test run: metered or unmetered. If the parameter is not specified, the default value is metered.

If you have unmetered device slots, you must set this to unmetered to use them. Otherwise, the run is counted toward metered device minutes.

skipAppResign :: Maybe Bool

When set to true, for private devices, Device Farm does not sign your app again. For public devices, Device Farm always signs your apps again.

For more information about how Device Farm re-signs your apps, see Do you modify my app? in the AWS Device Farm FAQs.

status :: Maybe ExecutionStatus

The run's status.

Allowed values include:

PENDING

PENDING_CONCURRENCY

PENDING_DEVICE

PROCESSING

SCHEDULING

PREPARING

RUNNING

COMPLETED

STOPPING

customerArtifactPaths :: Maybe CustomerArtifactPaths

Output CustomerArtifactPaths object for the test run.

eventCount :: Maybe Int

For fuzz tests, this is the number of events, between 1 and 10000, that the UI fuzz test should perform.

counters :: Maybe Counters

The run's result counters.

platform :: Maybe DevicePlatform

The run's platform.

Allowed values include:

ANDROID

IOS

seed :: Maybe Int

For fuzz tests, this is a seed to use for randomizing the UI fuzz test. Using the same seed value between tests ensures identical event sequences.

radios :: Maybe Radios

Information about the radio states for the run.

arn :: Maybe Text

The run's ARN.

location :: Maybe Location

Information about the location that is used for the run.

created :: Maybe POSIX

When the run was created.

locale :: Maybe Text

Information about the locale that is used for the run.

testSpecArn :: Maybe Text

The ARN of the YAML-formatted test specification for the run.

stopped :: Maybe POSIX

The run's stop time.

result :: Maybe ExecutionResult

The run's result.

Allowed values include:

PENDING

PASSED

WARNED

FAILED

SKIPPED

ERRORED

STOPPED

jobTimeoutMinutes :: Maybe Int

The number of minutes the job executes before it times out.

completedJobs :: Maybe Int

The total number of completed jobs.

resultCode :: Maybe ExecutionResultCode

Supporting field for the result field. Set only if result is SKIPPED. PARSING_FAILED if the result is skipped because of test package parsing failure.

name :: Maybe Text

The run's name.

appUpload :: Maybe Text

An app to upload or that has been uploaded.

parsingResultUrl :: Maybe Text

Read-only URL for an object in an S3 bucket where you can get the parsing results of the test package. If the test package doesn't parse, the reason why it doesn't parse appears in the file that this URL points to.

networkProfile :: Maybe NetworkProfile

The network profile being used for a test run.

deviceMinutes :: Maybe DeviceMinutes

Represents the total (metered or unmetered) minutes used by the test run.

type' :: Maybe TestType

The run's type.

Must be one of the following values:

BUILTIN_FUZZ

BUILTIN_EXPLORER

For Android, an app explorer that traverses an Android app, interacting with it and capturing screenshots at the same time.

APPIUM_JAVA_JUNIT

APPIUM_JAVA_TESTNG

APPIUM_PYTHON

APPIUM_NODE

APPIUM_RUBY

APPIUM_WEB_JAVA_JUNIT

APPIUM_WEB_JAVA_TESTNG

APPIUM_WEB_PYTHON

APPIUM_WEB_NODE

APPIUM_WEB_RUBY

CALABASH

INSTRUMENTATION

UIAUTOMATION

UIAUTOMATOR

XCTEST

XCTEST_UI

message :: Maybe Text

A message about the run's result.

webUrl :: Maybe Text

The Device Farm console URL for the recording of the run.

totalJobs :: Maybe Int

The total number of jobs for the run.

devicePoolArn :: Maybe Text

The ARN of the device pool for the run.

started :: Maybe POSIX

The run's start time.

deviceSelectionResult :: Maybe DeviceSelectionResult

The results of a device filter used to select the devices for a test run.

newRun :: Run

Create a value of Run with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:billingMethod:Run', run_billingMethod - Specifies the billing method for a test run: metered or unmetered. If the parameter is not specified, the default value is metered.

If you have unmetered device slots, you must set this to unmetered to use them. Otherwise, the run is counted toward metered device minutes.

$sel:skipAppResign:Run', run_skipAppResign - When set to true, for private devices, Device Farm does not sign your app again. For public devices, Device Farm always signs your apps again.

For more information about how Device Farm re-signs your apps, see Do you modify my app? in the AWS Device Farm FAQs.

$sel:status:Run', run_status - The run's status.

Allowed values include:

PENDING

PENDING_CONCURRENCY

PENDING_DEVICE

PROCESSING

SCHEDULING

PREPARING

RUNNING

COMPLETED

STOPPING

$sel:customerArtifactPaths:Run', run_customerArtifactPaths - Output CustomerArtifactPaths object for the test run.

$sel:eventCount:Run', run_eventCount - For fuzz tests, this is the number of events, between 1 and 10000, that the UI fuzz test should perform.

$sel:counters:Run', run_counters - The run's result counters.

$sel:platform:Run', run_platform - The run's platform.

Allowed values include:

ANDROID

IOS

$sel:seed:Run', run_seed - For fuzz tests, this is a seed to use for randomizing the UI fuzz test. Using the same seed value between tests ensures identical event sequences.

$sel:radios:Run', run_radios - Information about the radio states for the run.

$sel:arn:Run', run_arn - The run's ARN.

$sel:location:Run', run_location - Information about the location that is used for the run.

$sel:created:Run', run_created - When the run was created.

$sel:locale:Run', run_locale - Information about the locale that is used for the run.

$sel:testSpecArn:Run', run_testSpecArn - The ARN of the YAML-formatted test specification for the run.

$sel:stopped:Run', run_stopped - The run's stop time.

$sel:result:Run', run_result - The run's result.

Allowed values include:

PENDING

PASSED

WARNED

FAILED

SKIPPED

ERRORED

STOPPED

$sel:jobTimeoutMinutes:Run', run_jobTimeoutMinutes - The number of minutes the job executes before it times out.

$sel:completedJobs:Run', run_completedJobs - The total number of completed jobs.

$sel:resultCode:Run', run_resultCode - Supporting field for the result field. Set only if result is SKIPPED. PARSING_FAILED if the result is skipped because of test package parsing failure.

$sel:name:Run', run_name - The run's name.

$sel:appUpload:Run', run_appUpload - An app to upload or that has been uploaded.

$sel:parsingResultUrl:Run', run_parsingResultUrl - Read-only URL for an object in an S3 bucket where you can get the parsing results of the test package. If the test package doesn't parse, the reason why it doesn't parse appears in the file that this URL points to.

$sel:networkProfile:Run', run_networkProfile - The network profile being used for a test run.

$sel:deviceMinutes:Run', run_deviceMinutes - Represents the total (metered or unmetered) minutes used by the test run.

$sel:type':Run', run_type - The run's type.

Must be one of the following values:

BUILTIN_FUZZ

BUILTIN_EXPLORER

For Android, an app explorer that traverses an Android app, interacting with it and capturing screenshots at the same time.

APPIUM_JAVA_JUNIT

APPIUM_JAVA_TESTNG

APPIUM_PYTHON

APPIUM_NODE

APPIUM_RUBY

APPIUM_WEB_JAVA_JUNIT

APPIUM_WEB_JAVA_TESTNG

APPIUM_WEB_PYTHON

APPIUM_WEB_NODE

APPIUM_WEB_RUBY

CALABASH

INSTRUMENTATION

UIAUTOMATION

UIAUTOMATOR

XCTEST

XCTEST_UI

$sel:message:Run', run_message - A message about the run's result.

$sel:webUrl:Run', run_webUrl - The Device Farm console URL for the recording of the run.

$sel:totalJobs:Run', run_totalJobs - The total number of jobs for the run.

$sel:devicePoolArn:Run', run_devicePoolArn - The ARN of the device pool for the run.

$sel:started:Run', run_started - The run's start time.

$sel:deviceSelectionResult:Run', run_deviceSelectionResult - The results of a device filter used to select the devices for a test run.

run_billingMethod :: Lens' Run (Maybe BillingMethod)

Specifies the billing method for a test run: metered or unmetered. If the parameter is not specified, the default value is metered.

If you have unmetered device slots, you must set this to unmetered to use them. Otherwise, the run is counted toward metered device minutes.

run_skipAppResign :: Lens' Run (Maybe Bool)

When set to true, for private devices, Device Farm does not sign your app again. For public devices, Device Farm always signs your apps again.

For more information about how Device Farm re-signs your apps, see Do you modify my app? in the AWS Device Farm FAQs.

run_status :: Lens' Run (Maybe ExecutionStatus)

The run's status.

Allowed values include:

PENDING

PENDING_CONCURRENCY

PENDING_DEVICE

PROCESSING

SCHEDULING

PREPARING

RUNNING

COMPLETED

STOPPING

run_customerArtifactPaths :: Lens' Run (Maybe CustomerArtifactPaths)

Output CustomerArtifactPaths object for the test run.

run_eventCount :: Lens' Run (Maybe Int)

For fuzz tests, this is the number of events, between 1 and 10000, that the UI fuzz test should perform.

run_counters :: Lens' Run (Maybe Counters)

The run's result counters.

run_platform :: Lens' Run (Maybe DevicePlatform)

The run's platform.

Allowed values include:

ANDROID

IOS

run_seed :: Lens' Run (Maybe Int)

For fuzz tests, this is a seed to use for randomizing the UI fuzz test. Using the same seed value between tests ensures identical event sequences.

run_radios :: Lens' Run (Maybe Radios)

Information about the radio states for the run.

run_arn :: Lens' Run (Maybe Text)

The run's ARN.

run_location :: Lens' Run (Maybe Location)

Information about the location that is used for the run.

run_created :: Lens' Run (Maybe UTCTime)

When the run was created.

run_locale :: Lens' Run (Maybe Text)

Information about the locale that is used for the run.

run_testSpecArn :: Lens' Run (Maybe Text)

The ARN of the YAML-formatted test specification for the run.

run_stopped :: Lens' Run (Maybe UTCTime)

The run's stop time.

run_result :: Lens' Run (Maybe ExecutionResult)

The run's result.

Allowed values include:

PENDING

PASSED

WARNED

FAILED

SKIPPED

ERRORED

STOPPED

run_jobTimeoutMinutes :: Lens' Run (Maybe Int)

The number of minutes the job executes before it times out.

run_completedJobs :: Lens' Run (Maybe Int)

The total number of completed jobs.

run_resultCode :: Lens' Run (Maybe ExecutionResultCode)

Supporting field for the result field. Set only if result is SKIPPED. PARSING_FAILED if the result is skipped because of test package parsing failure.

run_name :: Lens' Run (Maybe Text)

The run's name.

run_appUpload :: Lens' Run (Maybe Text)

An app to upload or that has been uploaded.

run_parsingResultUrl :: Lens' Run (Maybe Text)

Read-only URL for an object in an S3 bucket where you can get the parsing results of the test package. If the test package doesn't parse, the reason why it doesn't parse appears in the file that this URL points to.

run_networkProfile :: Lens' Run (Maybe NetworkProfile)

The network profile being used for a test run.

run_deviceMinutes :: Lens' Run (Maybe DeviceMinutes)

Represents the total (metered or unmetered) minutes used by the test run.

run_type :: Lens' Run (Maybe TestType)

The run's type.

Must be one of the following values:

BUILTIN_FUZZ

BUILTIN_EXPLORER

For Android, an app explorer that traverses an Android app, interacting with it and capturing screenshots at the same time.

APPIUM_JAVA_JUNIT

APPIUM_JAVA_TESTNG

APPIUM_PYTHON

APPIUM_NODE

APPIUM_RUBY

APPIUM_WEB_JAVA_JUNIT

APPIUM_WEB_JAVA_TESTNG

APPIUM_WEB_PYTHON

APPIUM_WEB_NODE

APPIUM_WEB_RUBY

CALABASH

INSTRUMENTATION

UIAUTOMATION

UIAUTOMATOR

XCTEST

XCTEST_UI

run_message :: Lens' Run (Maybe Text)

A message about the run's result.

run_webUrl :: Lens' Run (Maybe Text)

The Device Farm console URL for the recording of the run.

run_totalJobs :: Lens' Run (Maybe Int)

The total number of jobs for the run.

run_devicePoolArn :: Lens' Run (Maybe Text)

The ARN of the device pool for the run.

run_started :: Lens' Run (Maybe UTCTime)

The run's start time.

run_deviceSelectionResult :: Lens' Run (Maybe DeviceSelectionResult)

The results of a device filter used to select the devices for a test run.