Network.AWS.DeviceFarm.Types.ScheduleRunConfiguration

Description

Documentation

data ScheduleRunConfiguration

Represents the settings for a run. Includes things like location, radio states, auxiliary apps, and network profiles.

See: newScheduleRunConfiguration smart constructor.

Constructors

ScheduleRunConfiguration'

Fields

auxiliaryApps :: Maybe [Text]

A list of upload ARNs for app packages to be installed with your app.

locale :: Maybe Text

Information about the locale that is used for the run.

radios :: Maybe Radios

Information about the radio states for the run.

billingMethod :: Maybe BillingMethod

Specifies the billing method for a test run: metered or unmetered. If the parameter is not specified, the default value is metered.

If you have purchased unmetered device slots, you must set this parameter to unmetered to make use of them. Otherwise, your run counts against your metered time.

customerArtifactPaths :: Maybe CustomerArtifactPaths

Input CustomerArtifactPaths object for the scheduled run configuration.

vpceConfigurationArns :: Maybe [Text]

An array of ARNs for your VPC endpoint configurations.

networkProfileArn :: Maybe Text

Reserved for internal use.

extraDataPackageArn :: Maybe Text

The ARN of the extra data for the run. The extra data is a .zip file that AWS Device Farm extracts to external data for Android or the app's sandbox for iOS.

location :: Maybe Location

Information about the location that is used for the run.

newScheduleRunConfiguration :: ScheduleRunConfiguration

Create a value of ScheduleRunConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:auxiliaryApps:ScheduleRunConfiguration', scheduleRunConfiguration_auxiliaryApps - A list of upload ARNs for app packages to be installed with your app.

$sel:locale:ScheduleRunConfiguration', scheduleRunConfiguration_locale - Information about the locale that is used for the run.

$sel:radios:ScheduleRunConfiguration', scheduleRunConfiguration_radios - Information about the radio states for the run.

$sel:billingMethod:ScheduleRunConfiguration', scheduleRunConfiguration_billingMethod - Specifies the billing method for a test run: metered or unmetered. If the parameter is not specified, the default value is metered.

If you have purchased unmetered device slots, you must set this parameter to unmetered to make use of them. Otherwise, your run counts against your metered time.

$sel:customerArtifactPaths:ScheduleRunConfiguration', scheduleRunConfiguration_customerArtifactPaths - Input CustomerArtifactPaths object for the scheduled run configuration.

$sel:vpceConfigurationArns:ScheduleRunConfiguration', scheduleRunConfiguration_vpceConfigurationArns - An array of ARNs for your VPC endpoint configurations.

$sel:networkProfileArn:ScheduleRunConfiguration', scheduleRunConfiguration_networkProfileArn - Reserved for internal use.

$sel:extraDataPackageArn:ScheduleRunConfiguration', scheduleRunConfiguration_extraDataPackageArn - The ARN of the extra data for the run. The extra data is a .zip file that AWS Device Farm extracts to external data for Android or the app's sandbox for iOS.

$sel:location:ScheduleRunConfiguration', scheduleRunConfiguration_location - Information about the location that is used for the run.

scheduleRunConfiguration_auxiliaryApps :: Lens' ScheduleRunConfiguration (Maybe [Text])

A list of upload ARNs for app packages to be installed with your app.

scheduleRunConfiguration_locale :: Lens' ScheduleRunConfiguration (Maybe Text)

Information about the locale that is used for the run.

scheduleRunConfiguration_radios :: Lens' ScheduleRunConfiguration (Maybe Radios)

Information about the radio states for the run.

scheduleRunConfiguration_billingMethod :: Lens' ScheduleRunConfiguration (Maybe BillingMethod)

Specifies the billing method for a test run: metered or unmetered. If the parameter is not specified, the default value is metered.

If you have purchased unmetered device slots, you must set this parameter to unmetered to make use of them. Otherwise, your run counts against your metered time.

scheduleRunConfiguration_customerArtifactPaths :: Lens' ScheduleRunConfiguration (Maybe CustomerArtifactPaths)

Input CustomerArtifactPaths object for the scheduled run configuration.

scheduleRunConfiguration_vpceConfigurationArns :: Lens' ScheduleRunConfiguration (Maybe [Text])

An array of ARNs for your VPC endpoint configurations.

scheduleRunConfiguration_networkProfileArn :: Lens' ScheduleRunConfiguration (Maybe Text)

Reserved for internal use.

scheduleRunConfiguration_extraDataPackageArn :: Lens' ScheduleRunConfiguration (Maybe Text)

The ARN of the extra data for the run. The extra data is a .zip file that AWS Device Farm extracts to external data for Android or the app's sandbox for iOS.

scheduleRunConfiguration_location :: Lens' ScheduleRunConfiguration (Maybe Location)

Information about the location that is used for the run.