Network.AWS.DeviceFarm.Types.TestGridSessionArtifact

Description

Documentation

data TestGridSessionArtifact

Artifacts are video and other files that are produced in the process of running a browser in an automated context.

Video elements might be broken up into multiple artifacts as they grow in size during creation.

See: newTestGridSessionArtifact smart constructor.

Constructors

TestGridSessionArtifact'

Fields

filename :: Maybe Text

The file name of the artifact.

url :: Maybe (Sensitive Text)

A semi-stable URL to the content of the object.

type' :: Maybe TestGridSessionArtifactType

The kind of artifact.

newTestGridSessionArtifact :: TestGridSessionArtifact

Create a value of TestGridSessionArtifact with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filename:TestGridSessionArtifact', testGridSessionArtifact_filename - The file name of the artifact.

$sel:url:TestGridSessionArtifact', testGridSessionArtifact_url - A semi-stable URL to the content of the object.

$sel:type':TestGridSessionArtifact', testGridSessionArtifact_type - The kind of artifact.

testGridSessionArtifact_filename :: Lens' TestGridSessionArtifact (Maybe Text)

The file name of the artifact.

testGridSessionArtifact_url :: Lens' TestGridSessionArtifact (Maybe Text)

A semi-stable URL to the content of the object.

testGridSessionArtifact_type :: Lens' TestGridSessionArtifact (Maybe TestGridSessionArtifactType)

The kind of artifact.