Network.AWS.DeviceFarm.Types.Upload

Description

Documentation

data Upload

An app or a set of one or more tests to upload or that have been uploaded.

See: newUpload smart constructor.

Constructors

Upload'

Fields

status :: Maybe UploadStatus

The upload's status.

Must be one of the following values:

FAILED

INITIALIZED

PROCESSING

SUCCEEDED

contentType :: Maybe Text

The upload's content type (for example, application/octet-stream).

message :: Maybe Text

A message about the upload's result.

category :: Maybe UploadCategory

The upload's category. Allowed values include:

CURATED: An upload managed by AWS Device Farm.

PRIVATE: An upload managed by the AWS Device Farm customer.

arn :: Maybe Text

The upload's ARN.

metadata :: Maybe Text

The upload's metadata. For example, for Android, this contains information that is parsed from the manifest and is displayed in the AWS Device Farm console after the associated app is uploaded.

name :: Maybe Text

The upload's file name.

url :: Maybe (Sensitive Text)

The presigned Amazon S3 URL that was used to store a file using a PUT request.

type' :: Maybe UploadType

The upload's type.

Must be one of the following values:

ANDROID_APP

IOS_APP

WEB_APP

EXTERNAL_DATA

APPIUM_JAVA_JUNIT_TEST_PACKAGE

APPIUM_JAVA_TESTNG_TEST_PACKAGE

APPIUM_PYTHON_TEST_PACKAGE

APPIUM_NODE_TEST_PACKAGE

APPIUM_RUBY_TEST_PACKAGE

APPIUM_WEB_JAVA_JUNIT_TEST_PACKAGE

APPIUM_WEB_JAVA_TESTNG_TEST_PACKAGE

APPIUM_WEB_PYTHON_TEST_PACKAGE

APPIUM_WEB_NODE_TEST_PACKAGE

APPIUM_WEB_RUBY_TEST_PACKAGE

CALABASH_TEST_PACKAGE

INSTRUMENTATION_TEST_PACKAGE

UIAUTOMATION_TEST_PACKAGE

UIAUTOMATOR_TEST_PACKAGE

XCTEST_TEST_PACKAGE

XCTEST_UI_TEST_PACKAGE

APPIUM_JAVA_JUNIT_TEST_SPEC

APPIUM_JAVA_TESTNG_TEST_SPEC

APPIUM_PYTHON_TEST_SPEC

APPIUM_NODE_TEST_SPEC

APPIUM_RUBY_TEST_SPEC

APPIUM_WEB_JAVA_JUNIT_TEST_SPEC

APPIUM_WEB_JAVA_TESTNG_TEST_SPEC

APPIUM_WEB_PYTHON_TEST_SPEC

APPIUM_WEB_NODE_TEST_SPEC

APPIUM_WEB_RUBY_TEST_SPEC

INSTRUMENTATION_TEST_SPEC

XCTEST_UI_TEST_SPEC

created :: Maybe POSIX

When the upload was created.

newUpload :: Upload

Create a value of Upload with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Upload', upload_status - The upload's status.

Must be one of the following values:

FAILED

INITIALIZED

PROCESSING

SUCCEEDED

$sel:contentType:Upload', upload_contentType - The upload's content type (for example, application/octet-stream).

$sel:message:Upload', upload_message - A message about the upload's result.

$sel:category:Upload', upload_category - The upload's category. Allowed values include:

CURATED: An upload managed by AWS Device Farm.

PRIVATE: An upload managed by the AWS Device Farm customer.

$sel:arn:Upload', upload_arn - The upload's ARN.

$sel:metadata:Upload', upload_metadata - The upload's metadata. For example, for Android, this contains information that is parsed from the manifest and is displayed in the AWS Device Farm console after the associated app is uploaded.

$sel:name:Upload', upload_name - The upload's file name.

$sel:url:Upload', upload_url - The presigned Amazon S3 URL that was used to store a file using a PUT request.

$sel:type':Upload', upload_type - The upload's type.

Must be one of the following values:

ANDROID_APP

IOS_APP

WEB_APP

EXTERNAL_DATA

APPIUM_JAVA_JUNIT_TEST_PACKAGE

APPIUM_JAVA_TESTNG_TEST_PACKAGE

APPIUM_PYTHON_TEST_PACKAGE

APPIUM_NODE_TEST_PACKAGE

APPIUM_RUBY_TEST_PACKAGE

APPIUM_WEB_JAVA_JUNIT_TEST_PACKAGE

APPIUM_WEB_JAVA_TESTNG_TEST_PACKAGE

APPIUM_WEB_PYTHON_TEST_PACKAGE

APPIUM_WEB_NODE_TEST_PACKAGE

APPIUM_WEB_RUBY_TEST_PACKAGE

CALABASH_TEST_PACKAGE

INSTRUMENTATION_TEST_PACKAGE

UIAUTOMATION_TEST_PACKAGE

UIAUTOMATOR_TEST_PACKAGE

XCTEST_TEST_PACKAGE

XCTEST_UI_TEST_PACKAGE

APPIUM_JAVA_JUNIT_TEST_SPEC

APPIUM_JAVA_TESTNG_TEST_SPEC

APPIUM_PYTHON_TEST_SPEC

APPIUM_NODE_TEST_SPEC

APPIUM_RUBY_TEST_SPEC

APPIUM_WEB_JAVA_JUNIT_TEST_SPEC

APPIUM_WEB_JAVA_TESTNG_TEST_SPEC

APPIUM_WEB_PYTHON_TEST_SPEC

APPIUM_WEB_NODE_TEST_SPEC

APPIUM_WEB_RUBY_TEST_SPEC

INSTRUMENTATION_TEST_SPEC

XCTEST_UI_TEST_SPEC

$sel:created:Upload', upload_created - When the upload was created.

upload_status :: Lens' Upload (Maybe UploadStatus)

The upload's status.

Must be one of the following values:

FAILED

INITIALIZED

PROCESSING

SUCCEEDED

upload_contentType :: Lens' Upload (Maybe Text)

The upload's content type (for example, application/octet-stream).

upload_message :: Lens' Upload (Maybe Text)

A message about the upload's result.

upload_category :: Lens' Upload (Maybe UploadCategory)

The upload's category. Allowed values include:

CURATED: An upload managed by AWS Device Farm.

PRIVATE: An upload managed by the AWS Device Farm customer.

upload_arn :: Lens' Upload (Maybe Text)

The upload's ARN.

upload_metadata :: Lens' Upload (Maybe Text)

The upload's metadata. For example, for Android, this contains information that is parsed from the manifest and is displayed in the AWS Device Farm console after the associated app is uploaded.

upload_name :: Lens' Upload (Maybe Text)

The upload's file name.

upload_url :: Lens' Upload (Maybe Text)

The presigned Amazon S3 URL that was used to store a file using a PUT request.

upload_type :: Lens' Upload (Maybe UploadType)

The upload's type.

Must be one of the following values:

ANDROID_APP

IOS_APP

WEB_APP

EXTERNAL_DATA

APPIUM_JAVA_JUNIT_TEST_PACKAGE

APPIUM_JAVA_TESTNG_TEST_PACKAGE

APPIUM_PYTHON_TEST_PACKAGE

APPIUM_NODE_TEST_PACKAGE

APPIUM_RUBY_TEST_PACKAGE

APPIUM_WEB_JAVA_JUNIT_TEST_PACKAGE

APPIUM_WEB_JAVA_TESTNG_TEST_PACKAGE

APPIUM_WEB_PYTHON_TEST_PACKAGE

APPIUM_WEB_NODE_TEST_PACKAGE

APPIUM_WEB_RUBY_TEST_PACKAGE

CALABASH_TEST_PACKAGE

INSTRUMENTATION_TEST_PACKAGE

UIAUTOMATION_TEST_PACKAGE

UIAUTOMATOR_TEST_PACKAGE

XCTEST_TEST_PACKAGE

XCTEST_UI_TEST_PACKAGE

APPIUM_JAVA_JUNIT_TEST_SPEC

APPIUM_JAVA_TESTNG_TEST_SPEC

APPIUM_PYTHON_TEST_SPEC

APPIUM_NODE_TEST_SPEC

APPIUM_RUBY_TEST_SPEC

APPIUM_WEB_JAVA_JUNIT_TEST_SPEC

APPIUM_WEB_JAVA_TESTNG_TEST_SPEC

APPIUM_WEB_PYTHON_TEST_SPEC

APPIUM_WEB_NODE_TEST_SPEC

APPIUM_WEB_RUBY_TEST_SPEC

INSTRUMENTATION_TEST_SPEC

XCTEST_UI_TEST_SPEC

upload_created :: Lens' Upload (Maybe UTCTime)

When the upload was created.