Network.AWS.DeviceFarm.Types.VPCEConfiguration

Description

Documentation

data VPCEConfiguration

Represents an Amazon Virtual Private Cloud (VPC) endpoint configuration.

See: newVPCEConfiguration smart constructor.

Constructors

VPCEConfiguration'

Fields

vpceConfigurationName :: Maybe Text

The friendly name you give to your VPC endpoint configuration to manage your configurations more easily.

vpceConfigurationDescription :: Maybe Text

An optional description that provides details about your VPC endpoint configuration.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the VPC endpoint configuration.

serviceDnsName :: Maybe Text

The DNS name that maps to the private IP address of the service you want to access.

vpceServiceName :: Maybe Text

The name of the VPC endpoint service running in your AWS account that you want Device Farm to test.

newVPCEConfiguration :: VPCEConfiguration

Create a value of VPCEConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpceConfigurationName:VPCEConfiguration', vPCEConfiguration_vpceConfigurationName - The friendly name you give to your VPC endpoint configuration to manage your configurations more easily.

$sel:vpceConfigurationDescription:VPCEConfiguration', vPCEConfiguration_vpceConfigurationDescription - An optional description that provides details about your VPC endpoint configuration.

$sel:arn:VPCEConfiguration', vPCEConfiguration_arn - The Amazon Resource Name (ARN) of the VPC endpoint configuration.

$sel:serviceDnsName:VPCEConfiguration', vPCEConfiguration_serviceDnsName - The DNS name that maps to the private IP address of the service you want to access.

$sel:vpceServiceName:VPCEConfiguration', vPCEConfiguration_vpceServiceName - The name of the VPC endpoint service running in your AWS account that you want Device Farm to test.

vPCEConfiguration_vpceConfigurationName :: Lens' VPCEConfiguration (Maybe Text)

The friendly name you give to your VPC endpoint configuration to manage your configurations more easily.

vPCEConfiguration_vpceConfigurationDescription :: Lens' VPCEConfiguration (Maybe Text)

An optional description that provides details about your VPC endpoint configuration.

vPCEConfiguration_arn :: Lens' VPCEConfiguration (Maybe Text)

The Amazon Resource Name (ARN) of the VPC endpoint configuration.

vPCEConfiguration_serviceDnsName :: Lens' VPCEConfiguration (Maybe Text)

The DNS name that maps to the private IP address of the service you want to access.

vPCEConfiguration_vpceServiceName :: Lens' VPCEConfiguration (Maybe Text)

The name of the VPC endpoint service running in your AWS account that you want Device Farm to test.