Network.AWS.DeviceFarm.UpdateNetworkProfile

Description

Updates the network profile.

Creating a Request

data UpdateNetworkProfile

See: newUpdateNetworkProfile smart constructor.

Constructors

UpdateNetworkProfile'

Fields

uplinkJitterMs :: Maybe Integer

Time variation in the delay of received packets in milliseconds as an integer from 0 to 2000.

downlinkBandwidthBits :: Maybe Integer

The data throughput rate in bits per second, as an integer from 0 to 104857600.

downlinkDelayMs :: Maybe Integer

Delay time for all packets to destination in milliseconds as an integer from 0 to 2000.

downlinkJitterMs :: Maybe Integer

Time variation in the delay of received packets in milliseconds as an integer from 0 to 2000.

uplinkLossPercent :: Maybe Natural

Proportion of transmitted packets that fail to arrive from 0 to 100 percent.

downlinkLossPercent :: Maybe Natural

Proportion of received packets that fail to arrive from 0 to 100 percent.

name :: Maybe Text

The name of the network profile about which you are returning information.

uplinkBandwidthBits :: Maybe Integer

The data throughput rate in bits per second, as an integer from 0 to 104857600.

uplinkDelayMs :: Maybe Integer

Delay time for all packets to destination in milliseconds as an integer from 0 to 2000.

description :: Maybe Text

The description of the network profile about which you are returning information.

type' :: Maybe NetworkProfileType

The type of network profile to return information about. Valid values are listed here.

arn :: Text

The Amazon Resource Name (ARN) of the project for which you want to update network profile settings.

newUpdateNetworkProfile

Arguments

:: Text
$sel:arn:UpdateNetworkProfile'

-> UpdateNetworkProfile

Create a value of UpdateNetworkProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:uplinkJitterMs:UpdateNetworkProfile', updateNetworkProfile_uplinkJitterMs - Time variation in the delay of received packets in milliseconds as an integer from 0 to 2000.

$sel:downlinkBandwidthBits:UpdateNetworkProfile', updateNetworkProfile_downlinkBandwidthBits - The data throughput rate in bits per second, as an integer from 0 to 104857600.

$sel:downlinkDelayMs:UpdateNetworkProfile', updateNetworkProfile_downlinkDelayMs - Delay time for all packets to destination in milliseconds as an integer from 0 to 2000.

$sel:downlinkJitterMs:UpdateNetworkProfile', updateNetworkProfile_downlinkJitterMs - Time variation in the delay of received packets in milliseconds as an integer from 0 to 2000.

$sel:uplinkLossPercent:UpdateNetworkProfile', updateNetworkProfile_uplinkLossPercent - Proportion of transmitted packets that fail to arrive from 0 to 100 percent.

$sel:downlinkLossPercent:UpdateNetworkProfile', updateNetworkProfile_downlinkLossPercent - Proportion of received packets that fail to arrive from 0 to 100 percent.

$sel:name:UpdateNetworkProfile', updateNetworkProfile_name - The name of the network profile about which you are returning information.

$sel:uplinkBandwidthBits:UpdateNetworkProfile', updateNetworkProfile_uplinkBandwidthBits - The data throughput rate in bits per second, as an integer from 0 to 104857600.

$sel:uplinkDelayMs:UpdateNetworkProfile', updateNetworkProfile_uplinkDelayMs - Delay time for all packets to destination in milliseconds as an integer from 0 to 2000.

$sel:description:UpdateNetworkProfile', updateNetworkProfile_description - The description of the network profile about which you are returning information.

$sel:type':UpdateNetworkProfile', updateNetworkProfile_type - The type of network profile to return information about. Valid values are listed here.

$sel:arn:UpdateNetworkProfile', updateNetworkProfile_arn - The Amazon Resource Name (ARN) of the project for which you want to update network profile settings.

Request Lenses

updateNetworkProfile_uplinkJitterMs :: Lens' UpdateNetworkProfile (Maybe Integer)

Time variation in the delay of received packets in milliseconds as an integer from 0 to 2000.

updateNetworkProfile_downlinkBandwidthBits :: Lens' UpdateNetworkProfile (Maybe Integer)

The data throughput rate in bits per second, as an integer from 0 to 104857600.

updateNetworkProfile_downlinkDelayMs :: Lens' UpdateNetworkProfile (Maybe Integer)

Delay time for all packets to destination in milliseconds as an integer from 0 to 2000.

updateNetworkProfile_downlinkJitterMs :: Lens' UpdateNetworkProfile (Maybe Integer)

Time variation in the delay of received packets in milliseconds as an integer from 0 to 2000.

updateNetworkProfile_uplinkLossPercent :: Lens' UpdateNetworkProfile (Maybe Natural)

Proportion of transmitted packets that fail to arrive from 0 to 100 percent.

updateNetworkProfile_downlinkLossPercent :: Lens' UpdateNetworkProfile (Maybe Natural)

Proportion of received packets that fail to arrive from 0 to 100 percent.

updateNetworkProfile_name :: Lens' UpdateNetworkProfile (Maybe Text)

The name of the network profile about which you are returning information.

updateNetworkProfile_uplinkBandwidthBits :: Lens' UpdateNetworkProfile (Maybe Integer)

The data throughput rate in bits per second, as an integer from 0 to 104857600.

updateNetworkProfile_uplinkDelayMs :: Lens' UpdateNetworkProfile (Maybe Integer)

Delay time for all packets to destination in milliseconds as an integer from 0 to 2000.

updateNetworkProfile_description :: Lens' UpdateNetworkProfile (Maybe Text)

The description of the network profile about which you are returning information.

updateNetworkProfile_type :: Lens' UpdateNetworkProfile (Maybe NetworkProfileType)

The type of network profile to return information about. Valid values are listed here.

updateNetworkProfile_arn :: Lens' UpdateNetworkProfile Text

The Amazon Resource Name (ARN) of the project for which you want to update network profile settings.

Destructuring the Response

data UpdateNetworkProfileResponse

See: newUpdateNetworkProfileResponse smart constructor.

Constructors

UpdateNetworkProfileResponse'

Fields

networkProfile :: Maybe NetworkProfile

A list of the available network profiles.

httpStatus :: Int

The response's http status code.

newUpdateNetworkProfileResponse

Arguments

:: Int
$sel:httpStatus:UpdateNetworkProfileResponse'

-> UpdateNetworkProfileResponse

Create a value of UpdateNetworkProfileResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:networkProfile:UpdateNetworkProfileResponse', updateNetworkProfileResponse_networkProfile - A list of the available network profiles.

$sel:httpStatus:UpdateNetworkProfileResponse', updateNetworkProfileResponse_httpStatus - The response's http status code.

Response Lenses

updateNetworkProfileResponse_networkProfile :: Lens' UpdateNetworkProfileResponse (Maybe NetworkProfile)

A list of the available network profiles.

updateNetworkProfileResponse_httpStatus :: Lens' UpdateNetworkProfileResponse Int

The response's http status code.