Network.AWS.DeviceFarm.UpdateTestGridProject

Description

Change details of a project.

Creating a Request

data UpdateTestGridProject

See: newUpdateTestGridProject smart constructor.

Constructors

UpdateTestGridProject'

Fields

vpcConfig :: Maybe TestGridVpcConfig

The VPC security groups and subnets that are attached to a project.

name :: Maybe Text

Human-readable name for the project.

description :: Maybe Text

Human-readable description for the project.

projectArn :: Text

ARN of the project to update.

newUpdateTestGridProject

Arguments

:: Text
$sel:projectArn:UpdateTestGridProject'

-> UpdateTestGridProject

Create a value of UpdateTestGridProject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpcConfig:UpdateTestGridProject', updateTestGridProject_vpcConfig - The VPC security groups and subnets that are attached to a project.

$sel:name:UpdateTestGridProject', updateTestGridProject_name - Human-readable name for the project.

$sel:description:UpdateTestGridProject', updateTestGridProject_description - Human-readable description for the project.

$sel:projectArn:UpdateTestGridProject', updateTestGridProject_projectArn - ARN of the project to update.

Request Lenses

updateTestGridProject_vpcConfig :: Lens' UpdateTestGridProject (Maybe TestGridVpcConfig)

The VPC security groups and subnets that are attached to a project.

updateTestGridProject_name :: Lens' UpdateTestGridProject (Maybe Text)

Human-readable name for the project.

updateTestGridProject_description :: Lens' UpdateTestGridProject (Maybe Text)

Human-readable description for the project.

updateTestGridProject_projectArn :: Lens' UpdateTestGridProject Text

ARN of the project to update.

Destructuring the Response

data UpdateTestGridProjectResponse

See: newUpdateTestGridProjectResponse smart constructor.

Constructors

UpdateTestGridProjectResponse'

Fields

testGridProject :: Maybe TestGridProject

The project, including updated information.

httpStatus :: Int

The response's http status code.

newUpdateTestGridProjectResponse

Arguments

:: Int
$sel:httpStatus:UpdateTestGridProjectResponse'

-> UpdateTestGridProjectResponse

Create a value of UpdateTestGridProjectResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:testGridProject:UpdateTestGridProjectResponse', updateTestGridProjectResponse_testGridProject - The project, including updated information.

$sel:httpStatus:UpdateTestGridProjectResponse', updateTestGridProjectResponse_httpStatus - The response's http status code.

Response Lenses

updateTestGridProjectResponse_testGridProject :: Lens' UpdateTestGridProjectResponse (Maybe TestGridProject)

The project, including updated information.

updateTestGridProjectResponse_httpStatus :: Lens' UpdateTestGridProjectResponse Int

The response's http status code.