Network.AWS.DirectConnect.AllocatePrivateVirtualInterface

Description

Provisions a private virtual interface to be owned by the specified account.

Virtual interfaces created using this action must be confirmed by the owner using ConfirmPrivateVirtualInterface. Until then, the virtual interface is in the Confirming state and is not available to handle traffic.

Creating a Request

data AllocatePrivateVirtualInterface

See: newAllocatePrivateVirtualInterface smart constructor.

Constructors

AllocatePrivateVirtualInterface'

Fields

connectionId :: Text

The ID of the connection on which the private virtual interface is provisioned.

ownerAccount :: Text

The ID of the account that owns the virtual private interface.

newPrivateVirtualInterfaceAllocation' :: NewPrivateVirtualInterfaceAllocation

Information about the private virtual interface.

newAllocatePrivateVirtualInterface

Arguments

:: Text
$sel:connectionId:AllocatePrivateVirtualInterface'

-> Text
$sel:ownerAccount:AllocatePrivateVirtualInterface'

-> NewPrivateVirtualInterfaceAllocation
$sel:newPrivateVirtualInterfaceAllocation':AllocatePrivateVirtualInterface'

-> AllocatePrivateVirtualInterface

Create a value of AllocatePrivateVirtualInterface with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:connectionId:AllocatePrivateVirtualInterface', allocatePrivateVirtualInterface_connectionId - The ID of the connection on which the private virtual interface is provisioned.

$sel:ownerAccount:AllocatePrivateVirtualInterface', allocatePrivateVirtualInterface_ownerAccount - The ID of the account that owns the virtual private interface.

$sel:newPrivateVirtualInterfaceAllocation':AllocatePrivateVirtualInterface', allocatePrivateVirtualInterface_newPrivateVirtualInterfaceAllocation - Information about the private virtual interface.

Request Lenses

allocatePrivateVirtualInterface_connectionId :: Lens' AllocatePrivateVirtualInterface Text

The ID of the connection on which the private virtual interface is provisioned.

allocatePrivateVirtualInterface_ownerAccount :: Lens' AllocatePrivateVirtualInterface Text

The ID of the account that owns the virtual private interface.

allocatePrivateVirtualInterface_newPrivateVirtualInterfaceAllocation :: Lens' AllocatePrivateVirtualInterface NewPrivateVirtualInterfaceAllocation

Information about the private virtual interface.

Destructuring the Response

data VirtualInterface

Information about a virtual interface.

See: newVirtualInterface smart constructor.

Constructors

VirtualInterface'

Fields

authKey :: Maybe Text

The authentication key for BGP configuration. This string has a minimum length of 6 characters and and a maximun lenth of 80 characters.

bgpPeers :: Maybe [BGPPeer]

The BGP peers configured on this virtual interface.

virtualGatewayId :: Maybe Text

The ID of the virtual private gateway. Applies only to private virtual interfaces.

asn :: Maybe Int

The autonomous system (AS) number for Border Gateway Protocol (BGP) configuration.

The valid values are 1-2147483647.

awsDeviceV2 :: Maybe Text

The Direct Connect endpoint that terminates the physical connection.

connectionId :: Maybe Text

The ID of the connection.

awsLogicalDeviceId :: Maybe Text

The Direct Connect endpoint that terminates the logical connection. This device might be different than the device that terminates the physical connection.

customerRouterConfig :: Maybe Text

The customer router configuration.

jumboFrameCapable :: Maybe Bool

Indicates whether jumbo frames (9001 MTU) are supported.

routeFilterPrefixes :: Maybe [RouteFilterPrefix]

The routes to be advertised to the Amazon Web Services network in this Region. Applies to public virtual interfaces.

mtu :: Maybe Int

The maximum transmission unit (MTU), in bytes. The supported values are 1500 and 9001. The default value is 1500.

virtualInterfaceType :: Maybe Text

The type of virtual interface. The possible values are private and public.

amazonSideAsn :: Maybe Integer

The autonomous system number (ASN) for the Amazon side of the connection.

virtualInterfaceId :: Maybe Text

The ID of the virtual interface.

tags :: Maybe (NonEmpty Tag)

The tags associated with the virtual interface.

directConnectGatewayId :: Maybe Text

The ID of the Direct Connect gateway.

virtualInterfaceName :: Maybe Text

The name of the virtual interface assigned by the customer network. The name has a maximum of 100 characters. The following are valid characters: a-z, 0-9 and a hyphen (-).

virtualInterfaceState :: Maybe VirtualInterfaceState

The state of the virtual interface. The following are the possible values:

confirming: The creation of the virtual interface is pending confirmation from the virtual interface owner. If the owner of the virtual interface is different from the owner of the connection on which it is provisioned, then the virtual interface will remain in this state until it is confirmed by the virtual interface owner.

verifying: This state only applies to public virtual interfaces. Each public virtual interface needs validation before the virtual interface can be created.

pending: A virtual interface is in this state from the time that it is created until the virtual interface is ready to forward traffic.

available: A virtual interface that is able to forward traffic.

down: A virtual interface that is BGP down.

deleting: A virtual interface is in this state immediately after calling DeleteVirtualInterface until it can no longer forward traffic.

deleted: A virtual interface that cannot forward traffic.

rejected: The virtual interface owner has declined creation of the virtual interface. If a virtual interface in the Confirming state is deleted by the virtual interface owner, the virtual interface enters the Rejected state.

unknown: The state of the virtual interface is not available.

amazonAddress :: Maybe Text

The IP address assigned to the Amazon interface.

addressFamily :: Maybe AddressFamily

The address family for the BGP peer.

ownerAccount :: Maybe Text

The ID of the account that owns the virtual interface.

region :: Maybe Text

The Region where the virtual interface is located.

vlan :: Maybe Int

The ID of the VLAN.

location :: Maybe Text

The location of the connection.

customerAddress :: Maybe Text

The IP address assigned to the customer interface.

newVirtualInterface :: VirtualInterface

Create a value of VirtualInterface with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:authKey:VirtualInterface', virtualInterface_authKey - The authentication key for BGP configuration. This string has a minimum length of 6 characters and and a maximun lenth of 80 characters.

$sel:bgpPeers:VirtualInterface', virtualInterface_bgpPeers - The BGP peers configured on this virtual interface.

$sel:virtualGatewayId:VirtualInterface', virtualInterface_virtualGatewayId - The ID of the virtual private gateway. Applies only to private virtual interfaces.

$sel:asn:VirtualInterface', virtualInterface_asn - The autonomous system (AS) number for Border Gateway Protocol (BGP) configuration.

The valid values are 1-2147483647.

$sel:awsDeviceV2:VirtualInterface', virtualInterface_awsDeviceV2 - The Direct Connect endpoint that terminates the physical connection.

$sel:connectionId:VirtualInterface', virtualInterface_connectionId - The ID of the connection.

$sel:awsLogicalDeviceId:VirtualInterface', virtualInterface_awsLogicalDeviceId - The Direct Connect endpoint that terminates the logical connection. This device might be different than the device that terminates the physical connection.

$sel:customerRouterConfig:VirtualInterface', virtualInterface_customerRouterConfig - The customer router configuration.

$sel:jumboFrameCapable:VirtualInterface', virtualInterface_jumboFrameCapable - Indicates whether jumbo frames (9001 MTU) are supported.

$sel:routeFilterPrefixes:VirtualInterface', virtualInterface_routeFilterPrefixes - The routes to be advertised to the Amazon Web Services network in this Region. Applies to public virtual interfaces.

$sel:mtu:VirtualInterface', virtualInterface_mtu - The maximum transmission unit (MTU), in bytes. The supported values are 1500 and 9001. The default value is 1500.

$sel:virtualInterfaceType:VirtualInterface', virtualInterface_virtualInterfaceType - The type of virtual interface. The possible values are private and public.

$sel:amazonSideAsn:VirtualInterface', virtualInterface_amazonSideAsn - The autonomous system number (ASN) for the Amazon side of the connection.

$sel:virtualInterfaceId:VirtualInterface', virtualInterface_virtualInterfaceId - The ID of the virtual interface.

$sel:tags:VirtualInterface', virtualInterface_tags - The tags associated with the virtual interface.

$sel:directConnectGatewayId:VirtualInterface', virtualInterface_directConnectGatewayId - The ID of the Direct Connect gateway.

$sel:virtualInterfaceName:VirtualInterface', virtualInterface_virtualInterfaceName - The name of the virtual interface assigned by the customer network. The name has a maximum of 100 characters. The following are valid characters: a-z, 0-9 and a hyphen (-).

$sel:virtualInterfaceState:VirtualInterface', virtualInterface_virtualInterfaceState - The state of the virtual interface. The following are the possible values:

confirming: The creation of the virtual interface is pending confirmation from the virtual interface owner. If the owner of the virtual interface is different from the owner of the connection on which it is provisioned, then the virtual interface will remain in this state until it is confirmed by the virtual interface owner.

verifying: This state only applies to public virtual interfaces. Each public virtual interface needs validation before the virtual interface can be created.

pending: A virtual interface is in this state from the time that it is created until the virtual interface is ready to forward traffic.

available: A virtual interface that is able to forward traffic.

down: A virtual interface that is BGP down.

deleting: A virtual interface is in this state immediately after calling DeleteVirtualInterface until it can no longer forward traffic.

deleted: A virtual interface that cannot forward traffic.

rejected: The virtual interface owner has declined creation of the virtual interface. If a virtual interface in the Confirming state is deleted by the virtual interface owner, the virtual interface enters the Rejected state.

unknown: The state of the virtual interface is not available.

$sel:amazonAddress:VirtualInterface', virtualInterface_amazonAddress - The IP address assigned to the Amazon interface.

$sel:addressFamily:VirtualInterface', virtualInterface_addressFamily - The address family for the BGP peer.

$sel:ownerAccount:VirtualInterface', virtualInterface_ownerAccount - The ID of the account that owns the virtual interface.

$sel:region:VirtualInterface', virtualInterface_region - The Region where the virtual interface is located.

$sel:vlan:VirtualInterface', virtualInterface_vlan - The ID of the VLAN.

$sel:location:VirtualInterface', virtualInterface_location - The location of the connection.

$sel:customerAddress:VirtualInterface', virtualInterface_customerAddress - The IP address assigned to the customer interface.

Response Lenses

virtualInterface_authKey :: Lens' VirtualInterface (Maybe Text)

The authentication key for BGP configuration. This string has a minimum length of 6 characters and and a maximun lenth of 80 characters.

virtualInterface_bgpPeers :: Lens' VirtualInterface (Maybe [BGPPeer])

The BGP peers configured on this virtual interface.

virtualInterface_virtualGatewayId :: Lens' VirtualInterface (Maybe Text)

The ID of the virtual private gateway. Applies only to private virtual interfaces.

virtualInterface_asn :: Lens' VirtualInterface (Maybe Int)

The autonomous system (AS) number for Border Gateway Protocol (BGP) configuration.

The valid values are 1-2147483647.

virtualInterface_awsDeviceV2 :: Lens' VirtualInterface (Maybe Text)

The Direct Connect endpoint that terminates the physical connection.

virtualInterface_connectionId :: Lens' VirtualInterface (Maybe Text)

The ID of the connection.

virtualInterface_awsLogicalDeviceId :: Lens' VirtualInterface (Maybe Text)

The Direct Connect endpoint that terminates the logical connection. This device might be different than the device that terminates the physical connection.

virtualInterface_customerRouterConfig :: Lens' VirtualInterface (Maybe Text)

The customer router configuration.

virtualInterface_jumboFrameCapable :: Lens' VirtualInterface (Maybe Bool)

Indicates whether jumbo frames (9001 MTU) are supported.

virtualInterface_routeFilterPrefixes :: Lens' VirtualInterface (Maybe [RouteFilterPrefix])

The routes to be advertised to the Amazon Web Services network in this Region. Applies to public virtual interfaces.

virtualInterface_mtu :: Lens' VirtualInterface (Maybe Int)

The maximum transmission unit (MTU), in bytes. The supported values are 1500 and 9001. The default value is 1500.

virtualInterface_virtualInterfaceType :: Lens' VirtualInterface (Maybe Text)

The type of virtual interface. The possible values are private and public.

virtualInterface_amazonSideAsn :: Lens' VirtualInterface (Maybe Integer)

The autonomous system number (ASN) for the Amazon side of the connection.

virtualInterface_virtualInterfaceId :: Lens' VirtualInterface (Maybe Text)

The ID of the virtual interface.

virtualInterface_tags :: Lens' VirtualInterface (Maybe (NonEmpty Tag))

The tags associated with the virtual interface.

virtualInterface_directConnectGatewayId :: Lens' VirtualInterface (Maybe Text)

The ID of the Direct Connect gateway.

virtualInterface_virtualInterfaceName :: Lens' VirtualInterface (Maybe Text)

The name of the virtual interface assigned by the customer network. The name has a maximum of 100 characters. The following are valid characters: a-z, 0-9 and a hyphen (-).

virtualInterface_virtualInterfaceState :: Lens' VirtualInterface (Maybe VirtualInterfaceState)

The state of the virtual interface. The following are the possible values:

confirming: The creation of the virtual interface is pending confirmation from the virtual interface owner. If the owner of the virtual interface is different from the owner of the connection on which it is provisioned, then the virtual interface will remain in this state until it is confirmed by the virtual interface owner.

verifying: This state only applies to public virtual interfaces. Each public virtual interface needs validation before the virtual interface can be created.

pending: A virtual interface is in this state from the time that it is created until the virtual interface is ready to forward traffic.

available: A virtual interface that is able to forward traffic.

down: A virtual interface that is BGP down.

deleting: A virtual interface is in this state immediately after calling DeleteVirtualInterface until it can no longer forward traffic.

deleted: A virtual interface that cannot forward traffic.

rejected: The virtual interface owner has declined creation of the virtual interface. If a virtual interface in the Confirming state is deleted by the virtual interface owner, the virtual interface enters the Rejected state.

unknown: The state of the virtual interface is not available.

virtualInterface_amazonAddress :: Lens' VirtualInterface (Maybe Text)

The IP address assigned to the Amazon interface.

virtualInterface_addressFamily :: Lens' VirtualInterface (Maybe AddressFamily)

The address family for the BGP peer.

virtualInterface_ownerAccount :: Lens' VirtualInterface (Maybe Text)

The ID of the account that owns the virtual interface.

virtualInterface_region :: Lens' VirtualInterface (Maybe Text)

The Region where the virtual interface is located.

virtualInterface_vlan :: Lens' VirtualInterface (Maybe Int)

The ID of the VLAN.

virtualInterface_location :: Lens' VirtualInterface (Maybe Text)

The location of the connection.

virtualInterface_customerAddress :: Lens' VirtualInterface (Maybe Text)

The IP address assigned to the customer interface.