Network.AWS.DirectConnect.AllocateTransitVirtualInterface

Description

Provisions a transit virtual interface to be owned by the specified Amazon Web Services account. Use this type of interface to connect a transit gateway to your Direct Connect gateway.

The owner of a connection provisions a transit virtual interface to be owned by the specified Amazon Web Services account.

After you create a transit virtual interface, it must be confirmed by the owner using ConfirmTransitVirtualInterface. Until this step has been completed, the transit virtual interface is in the requested state and is not available to handle traffic.

Creating a Request

data AllocateTransitVirtualInterface

See: newAllocateTransitVirtualInterface smart constructor.

Constructors

AllocateTransitVirtualInterface'

Fields

connectionId :: Text

The ID of the connection on which the transit virtual interface is provisioned.

ownerAccount :: Text

The ID of the Amazon Web Services account that owns the transit virtual interface.

newTransitVirtualInterfaceAllocation' :: NewTransitVirtualInterfaceAllocation

Information about the transit virtual interface.

newAllocateTransitVirtualInterface

Arguments

:: Text
$sel:connectionId:AllocateTransitVirtualInterface'

-> Text
$sel:ownerAccount:AllocateTransitVirtualInterface'

-> NewTransitVirtualInterfaceAllocation
$sel:newTransitVirtualInterfaceAllocation':AllocateTransitVirtualInterface'

-> AllocateTransitVirtualInterface

Create a value of AllocateTransitVirtualInterface with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:connectionId:AllocateTransitVirtualInterface', allocateTransitVirtualInterface_connectionId - The ID of the connection on which the transit virtual interface is provisioned.

$sel:ownerAccount:AllocateTransitVirtualInterface', allocateTransitVirtualInterface_ownerAccount - The ID of the Amazon Web Services account that owns the transit virtual interface.

$sel:newTransitVirtualInterfaceAllocation':AllocateTransitVirtualInterface', allocateTransitVirtualInterface_newTransitVirtualInterfaceAllocation - Information about the transit virtual interface.

Request Lenses

allocateTransitVirtualInterface_connectionId :: Lens' AllocateTransitVirtualInterface Text

The ID of the connection on which the transit virtual interface is provisioned.

allocateTransitVirtualInterface_ownerAccount :: Lens' AllocateTransitVirtualInterface Text

The ID of the Amazon Web Services account that owns the transit virtual interface.

allocateTransitVirtualInterface_newTransitVirtualInterfaceAllocation :: Lens' AllocateTransitVirtualInterface NewTransitVirtualInterfaceAllocation

Information about the transit virtual interface.

Destructuring the Response

data AllocateTransitVirtualInterfaceResponse

See: newAllocateTransitVirtualInterfaceResponse smart constructor.

Constructors

AllocateTransitVirtualInterfaceResponse'

Fields

virtualInterface :: Maybe VirtualInterface

httpStatus :: Int

The response's http status code.

newAllocateTransitVirtualInterfaceResponse

Arguments

:: Int
$sel:httpStatus:AllocateTransitVirtualInterfaceResponse'

-> AllocateTransitVirtualInterfaceResponse

Create a value of AllocateTransitVirtualInterfaceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:virtualInterface:AllocateTransitVirtualInterfaceResponse', allocateTransitVirtualInterfaceResponse_virtualInterface - Undocumented member.

$sel:httpStatus:AllocateTransitVirtualInterfaceResponse', allocateTransitVirtualInterfaceResponse_httpStatus - The response's http status code.

Response Lenses

allocateTransitVirtualInterfaceResponse_virtualInterface :: Lens' AllocateTransitVirtualInterfaceResponse (Maybe VirtualInterface)

Undocumented member.

allocateTransitVirtualInterfaceResponse_httpStatus :: Lens' AllocateTransitVirtualInterfaceResponse Int

The response's http status code.