Network.AWS.DirectConnect.AssociateMacSecKey

Description

Associates a MAC Security (MACsec) Connection Key Name (CKN)/ Connectivity Association Key (CAK) pair with an Direct Connect dedicated connection.

You must supply either the secretARN, or the CKN/CAK (ckn and cak) pair in the request.

For information about MAC Security (MACsec) key considerations, see MACsec pre-shared CKN/CAK key considerations in the Direct Connect User Guide.

Creating a Request

data AssociateMacSecKey

See: newAssociateMacSecKey smart constructor.

Constructors

AssociateMacSecKey'

Fields

ckn :: Maybe Text

The MAC Security (MACsec) CKN to associate with the dedicated connection.

You can create the CKN/CAK pair using an industry standard tool.

The valid values are 64 hexadecimal characters (0-9, A-E).

If you use this request parameter, you must use the cak request parameter and not use the secretARN request parameter.

secretARN :: Maybe Text

The Amazon Resource Name (ARN) of the MAC Security (MACsec) secret key to associate with the dedicated connection.

You can use DescribeConnections or DescribeLags to retrieve the MAC Security (MACsec) secret key.

If you use this request parameter, you do not use the ckn and cak request parameters.

cak :: Maybe Text

The MAC Security (MACsec) CAK to associate with the dedicated connection.

You can create the CKN/CAK pair using an industry standard tool.

The valid values are 64 hexadecimal characters (0-9, A-E).

If you use this request parameter, you must use the ckn request parameter and not use the secretARN request parameter.

connectionId :: Text

The ID of the dedicated connection (dxcon-xxxx), or the ID of the LAG (dxlag-xxxx).

You can use DescribeConnections or DescribeLags to retrieve connection ID.

newAssociateMacSecKey

Arguments

:: Text
$sel:connectionId:AssociateMacSecKey'

-> AssociateMacSecKey

Create a value of AssociateMacSecKey with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ckn:AssociateMacSecKey', associateMacSecKey_ckn - The MAC Security (MACsec) CKN to associate with the dedicated connection.

You can create the CKN/CAK pair using an industry standard tool.

The valid values are 64 hexadecimal characters (0-9, A-E).

If you use this request parameter, you must use the cak request parameter and not use the secretARN request parameter.

$sel:secretARN:AssociateMacSecKey', associateMacSecKey_secretARN - The Amazon Resource Name (ARN) of the MAC Security (MACsec) secret key to associate with the dedicated connection.

You can use DescribeConnections or DescribeLags to retrieve the MAC Security (MACsec) secret key.

If you use this request parameter, you do not use the ckn and cak request parameters.

$sel:cak:AssociateMacSecKey', associateMacSecKey_cak - The MAC Security (MACsec) CAK to associate with the dedicated connection.

You can create the CKN/CAK pair using an industry standard tool.

The valid values are 64 hexadecimal characters (0-9, A-E).

If you use this request parameter, you must use the ckn request parameter and not use the secretARN request parameter.

$sel:connectionId:AssociateMacSecKey', associateMacSecKey_connectionId - The ID of the dedicated connection (dxcon-xxxx), or the ID of the LAG (dxlag-xxxx).

You can use DescribeConnections or DescribeLags to retrieve connection ID.

Request Lenses

associateMacSecKey_ckn :: Lens' AssociateMacSecKey (Maybe Text)

The MAC Security (MACsec) CKN to associate with the dedicated connection.

You can create the CKN/CAK pair using an industry standard tool.

The valid values are 64 hexadecimal characters (0-9, A-E).

If you use this request parameter, you must use the cak request parameter and not use the secretARN request parameter.

associateMacSecKey_secretARN :: Lens' AssociateMacSecKey (Maybe Text)

The Amazon Resource Name (ARN) of the MAC Security (MACsec) secret key to associate with the dedicated connection.

You can use DescribeConnections or DescribeLags to retrieve the MAC Security (MACsec) secret key.

If you use this request parameter, you do not use the ckn and cak request parameters.

associateMacSecKey_cak :: Lens' AssociateMacSecKey (Maybe Text)

The MAC Security (MACsec) CAK to associate with the dedicated connection.

You can create the CKN/CAK pair using an industry standard tool.

The valid values are 64 hexadecimal characters (0-9, A-E).

If you use this request parameter, you must use the ckn request parameter and not use the secretARN request parameter.

associateMacSecKey_connectionId :: Lens' AssociateMacSecKey Text

The ID of the dedicated connection (dxcon-xxxx), or the ID of the LAG (dxlag-xxxx).

You can use DescribeConnections or DescribeLags to retrieve connection ID.

Destructuring the Response

data AssociateMacSecKeyResponse

See: newAssociateMacSecKeyResponse smart constructor.

Constructors

AssociateMacSecKeyResponse'

Fields

macSecKeys :: Maybe [MacSecKey]

The MAC Security (MACsec) security keys associated with the dedicated connection.

connectionId :: Maybe Text

The ID of the dedicated connection (dxcon-xxxx), or the ID of the LAG (dxlag-xxxx).

httpStatus :: Int

The response's http status code.

newAssociateMacSecKeyResponse

Arguments

:: Int
$sel:httpStatus:AssociateMacSecKeyResponse'

-> AssociateMacSecKeyResponse

Create a value of AssociateMacSecKeyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:macSecKeys:AssociateMacSecKeyResponse', associateMacSecKeyResponse_macSecKeys - The MAC Security (MACsec) security keys associated with the dedicated connection.

$sel:connectionId:AssociateMacSecKey', associateMacSecKeyResponse_connectionId - The ID of the dedicated connection (dxcon-xxxx), or the ID of the LAG (dxlag-xxxx).

$sel:httpStatus:AssociateMacSecKeyResponse', associateMacSecKeyResponse_httpStatus - The response's http status code.

Response Lenses

associateMacSecKeyResponse_macSecKeys :: Lens' AssociateMacSecKeyResponse (Maybe [MacSecKey])

The MAC Security (MACsec) security keys associated with the dedicated connection.

associateMacSecKeyResponse_connectionId :: Lens' AssociateMacSecKeyResponse (Maybe Text)

The ID of the dedicated connection (dxcon-xxxx), or the ID of the LAG (dxlag-xxxx).

associateMacSecKeyResponse_httpStatus :: Lens' AssociateMacSecKeyResponse Int

The response's http status code.