Network.AWS.DirectConnect.ConfirmConnection

Description

Confirms the creation of the specified hosted connection on an interconnect.

Upon creation, the hosted connection is initially in the Ordering state, and remains in this state until the owner confirms creation of the hosted connection.

Creating a Request

data ConfirmConnection

See: newConfirmConnection smart constructor.

Constructors

ConfirmConnection'

Fields

connectionId :: Text

The ID of the hosted connection.

newConfirmConnection

Arguments

:: Text
$sel:connectionId:ConfirmConnection'

-> ConfirmConnection

Create a value of ConfirmConnection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:connectionId:ConfirmConnection', confirmConnection_connectionId - The ID of the hosted connection.

Request Lenses

confirmConnection_connectionId :: Lens' ConfirmConnection Text

The ID of the hosted connection.

Destructuring the Response

data ConfirmConnectionResponse

See: newConfirmConnectionResponse smart constructor.

Constructors

ConfirmConnectionResponse'

Fields

connectionState :: Maybe ConnectionState

The state of the connection. The following are the possible values:

ordering: The initial state of a hosted connection provisioned on an interconnect. The connection stays in the ordering state until the owner of the hosted connection confirms or declines the connection order.

requested: The initial state of a standard connection. The connection stays in the requested state until the Letter of Authorization (LOA) is sent to the customer.

pending: The connection has been approved and is being initialized.

available: The network link is up and the connection is ready for use.

down: The network link is down.

deleting: The connection is being deleted.

deleted: The connection has been deleted.

rejected: A hosted connection in the ordering state enters the rejected state if it is deleted by the customer.

unknown: The state of the connection is not available.

httpStatus :: Int

The response's http status code.

newConfirmConnectionResponse

Arguments

:: Int
$sel:httpStatus:ConfirmConnectionResponse'

-> ConfirmConnectionResponse

Create a value of ConfirmConnectionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:connectionState:ConfirmConnectionResponse', confirmConnectionResponse_connectionState - The state of the connection. The following are the possible values:

ordering: The initial state of a hosted connection provisioned on an interconnect. The connection stays in the ordering state until the owner of the hosted connection confirms or declines the connection order.

requested: The initial state of a standard connection. The connection stays in the requested state until the Letter of Authorization (LOA) is sent to the customer.

pending: The connection has been approved and is being initialized.

available: The network link is up and the connection is ready for use.

down: The network link is down.

deleting: The connection is being deleted.

deleted: The connection has been deleted.

rejected: A hosted connection in the ordering state enters the rejected state if it is deleted by the customer.

unknown: The state of the connection is not available.

$sel:httpStatus:ConfirmConnectionResponse', confirmConnectionResponse_httpStatus - The response's http status code.

Response Lenses

confirmConnectionResponse_connectionState :: Lens' ConfirmConnectionResponse (Maybe ConnectionState)

The state of the connection. The following are the possible values:

ordering: The initial state of a hosted connection provisioned on an interconnect. The connection stays in the ordering state until the owner of the hosted connection confirms or declines the connection order.

requested: The initial state of a standard connection. The connection stays in the requested state until the Letter of Authorization (LOA) is sent to the customer.

pending: The connection has been approved and is being initialized.

available: The network link is up and the connection is ready for use.

down: The network link is down.

deleting: The connection is being deleted.

deleted: The connection has been deleted.

rejected: A hosted connection in the ordering state enters the rejected state if it is deleted by the customer.

unknown: The state of the connection is not available.

confirmConnectionResponse_httpStatus :: Lens' ConfirmConnectionResponse Int

The response's http status code.