Network.AWS.DirectConnect.CreateBGPPeer

Description

Creates a BGP peer on the specified virtual interface.

You must create a BGP peer for the corresponding address family (IPv4/IPv6) in order to access Amazon Web Services resources that also use that address family.

If logical redundancy is not supported by the connection, interconnect, or LAG, the BGP peer cannot be in the same address family as an existing BGP peer on the virtual interface.

When creating a IPv6 BGP peer, omit the Amazon address and customer address. IPv6 addresses are automatically assigned from the Amazon pool of IPv6 addresses; you cannot specify custom IPv6 addresses.

For a public virtual interface, the Autonomous System Number (ASN) must be private or already on the allow list for the virtual interface.

Creating a Request

data CreateBGPPeer

See: newCreateBGPPeer smart constructor.

Constructors

CreateBGPPeer'

Fields

virtualInterfaceId :: Maybe Text

The ID of the virtual interface.

newBGPPeer' :: Maybe NewBGPPeer

Information about the BGP peer.

newCreateBGPPeer :: CreateBGPPeer

Create a value of CreateBGPPeer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:virtualInterfaceId:CreateBGPPeer', createBGPPeer_virtualInterfaceId - The ID of the virtual interface.

$sel:newBGPPeer':CreateBGPPeer', createBGPPeer_newBGPPeer - Information about the BGP peer.

Request Lenses

createBGPPeer_virtualInterfaceId :: Lens' CreateBGPPeer (Maybe Text)

The ID of the virtual interface.

createBGPPeer_newBGPPeer :: Lens' CreateBGPPeer (Maybe NewBGPPeer)

Information about the BGP peer.

Destructuring the Response

data CreateBGPPeerResponse

See: newCreateBGPPeerResponse smart constructor.

Constructors

CreateBGPPeerResponse'

Fields

virtualInterface :: Maybe VirtualInterface

The virtual interface.

httpStatus :: Int

The response's http status code.

newCreateBGPPeerResponse

Arguments

:: Int
$sel:httpStatus:CreateBGPPeerResponse'

-> CreateBGPPeerResponse

Create a value of CreateBGPPeerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:virtualInterface:CreateBGPPeerResponse', createBGPPeerResponse_virtualInterface - The virtual interface.

$sel:httpStatus:CreateBGPPeerResponse', createBGPPeerResponse_httpStatus - The response's http status code.

Response Lenses

createBGPPeerResponse_virtualInterface :: Lens' CreateBGPPeerResponse (Maybe VirtualInterface)

The virtual interface.

createBGPPeerResponse_httpStatus :: Lens' CreateBGPPeerResponse Int

The response's http status code.