Network.AWS.DirectConnect.DeleteLag

Description

Deletes the specified link aggregation group (LAG). You cannot delete a LAG if it has active virtual interfaces or hosted connections.

Creating a Request

data DeleteLag

See: newDeleteLag smart constructor.

Constructors

DeleteLag'

Fields

lagId :: Text

The ID of the LAG.

newDeleteLag

Arguments

:: Text
$sel:lagId:DeleteLag'

-> DeleteLag

Create a value of DeleteLag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lagId:DeleteLag', deleteLag_lagId - The ID of the LAG.

Request Lenses

deleteLag_lagId :: Lens' DeleteLag Text

The ID of the LAG.

Destructuring the Response

data Lag

Information about a link aggregation group (LAG).

See: newLag smart constructor.

Constructors

Lag'

Fields

lagId :: Maybe Text

The ID of the LAG.

macSecCapable :: Maybe Bool

Indicates whether the LAG supports MAC Security (MACsec).

connectionsBandwidth :: Maybe Text

The individual bandwidth of the physical connections bundled by the LAG. The possible values are 1Gbps and 10Gbps.

minimumLinks :: Maybe Int

The minimum number of physical dedicated connections that must be operational for the LAG itself to be operational.

lagName :: Maybe Text

The name of the LAG.

location :: Maybe Text

The location of the LAG.

connections :: Maybe [Connection]

The connections bundled by the LAG.

awsDevice :: Maybe Text

The Direct Connect endpoint that hosts the LAG.

hasLogicalRedundancy :: Maybe HasLogicalRedundancy

Indicates whether the LAG supports a secondary BGP peer in the same address family (IPv4/IPv6).

awsLogicalDeviceId :: Maybe Text

The Direct Connect endpoint that terminates the logical connection. This device might be different than the device that terminates the physical connection.

allowsHostedConnections :: Maybe Bool

Indicates whether the LAG can host other connections.

encryptionMode :: Maybe Text

The LAG MAC Security (MACsec) encryption mode.

The valid values are no_encrypt, should_encrypt, and must_encrypt.

numberOfConnections :: Maybe Int

The number of physical dedicated connections bundled by the LAG, up to a maximum of 10.

jumboFrameCapable :: Maybe Bool

Indicates whether jumbo frames (9001 MTU) are supported.

lagState :: Maybe LagState

The state of the LAG. The following are the possible values:

requested: The initial state of a LAG. The LAG stays in the requested state until the Letter of Authorization (LOA) is available.

pending: The LAG has been approved and is being initialized.

available: The network link is established and the LAG is ready for use.

down: The network link is down.

deleting: The LAG is being deleted.

deleted: The LAG is deleted.

unknown: The state of the LAG is not available.

ownerAccount :: Maybe Text

The ID of the account that owns the LAG.

region :: Maybe Text

The Region where the connection is located.

macSecKeys :: Maybe [MacSecKey]

The MAC Security (MACsec) security keys associated with the LAG.

providerName :: Maybe Text

The name of the service provider associated with the LAG.

awsDeviceV2 :: Maybe Text

The Direct Connect endpoint that hosts the LAG.

tags :: Maybe (NonEmpty Tag)

The tags associated with the LAG.

newLag :: Lag

Create a value of Lag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lagId:Lag', lag_lagId - The ID of the LAG.

$sel:macSecCapable:Lag', lag_macSecCapable - Indicates whether the LAG supports MAC Security (MACsec).

$sel:connectionsBandwidth:Lag', lag_connectionsBandwidth - The individual bandwidth of the physical connections bundled by the LAG. The possible values are 1Gbps and 10Gbps.

$sel:minimumLinks:Lag', lag_minimumLinks - The minimum number of physical dedicated connections that must be operational for the LAG itself to be operational.

$sel:lagName:Lag', lag_lagName - The name of the LAG.

$sel:location:Lag', lag_location - The location of the LAG.

$sel:connections:Lag', lag_connections - The connections bundled by the LAG.

$sel:awsDevice:Lag', lag_awsDevice - The Direct Connect endpoint that hosts the LAG.

$sel:hasLogicalRedundancy:Lag', lag_hasLogicalRedundancy - Indicates whether the LAG supports a secondary BGP peer in the same address family (IPv4/IPv6).

$sel:awsLogicalDeviceId:Lag', lag_awsLogicalDeviceId - The Direct Connect endpoint that terminates the logical connection. This device might be different than the device that terminates the physical connection.

$sel:allowsHostedConnections:Lag', lag_allowsHostedConnections - Indicates whether the LAG can host other connections.

$sel:encryptionMode:Lag', lag_encryptionMode - The LAG MAC Security (MACsec) encryption mode.

The valid values are no_encrypt, should_encrypt, and must_encrypt.

$sel:numberOfConnections:Lag', lag_numberOfConnections - The number of physical dedicated connections bundled by the LAG, up to a maximum of 10.

$sel:jumboFrameCapable:Lag', lag_jumboFrameCapable - Indicates whether jumbo frames (9001 MTU) are supported.

$sel:lagState:Lag', lag_lagState - The state of the LAG. The following are the possible values:

requested: The initial state of a LAG. The LAG stays in the requested state until the Letter of Authorization (LOA) is available.

pending: The LAG has been approved and is being initialized.

available: The network link is established and the LAG is ready for use.

down: The network link is down.

deleting: The LAG is being deleted.

deleted: The LAG is deleted.

unknown: The state of the LAG is not available.

$sel:ownerAccount:Lag', lag_ownerAccount - The ID of the account that owns the LAG.

$sel:region:Lag', lag_region - The Region where the connection is located.

$sel:macSecKeys:Lag', lag_macSecKeys - The MAC Security (MACsec) security keys associated with the LAG.

$sel:providerName:Lag', lag_providerName - The name of the service provider associated with the LAG.

$sel:awsDeviceV2:Lag', lag_awsDeviceV2 - The Direct Connect endpoint that hosts the LAG.

$sel:tags:Lag', lag_tags - The tags associated with the LAG.

Response Lenses

lag_lagId :: Lens' Lag (Maybe Text)

The ID of the LAG.

lag_macSecCapable :: Lens' Lag (Maybe Bool)

Indicates whether the LAG supports MAC Security (MACsec).

lag_connectionsBandwidth :: Lens' Lag (Maybe Text)

The individual bandwidth of the physical connections bundled by the LAG. The possible values are 1Gbps and 10Gbps.

lag_minimumLinks :: Lens' Lag (Maybe Int)

The minimum number of physical dedicated connections that must be operational for the LAG itself to be operational.

lag_lagName :: Lens' Lag (Maybe Text)

The name of the LAG.

lag_location :: Lens' Lag (Maybe Text)

The location of the LAG.

lag_connections :: Lens' Lag (Maybe [Connection])

The connections bundled by the LAG.

lag_awsDevice :: Lens' Lag (Maybe Text)

The Direct Connect endpoint that hosts the LAG.

lag_hasLogicalRedundancy :: Lens' Lag (Maybe HasLogicalRedundancy)

Indicates whether the LAG supports a secondary BGP peer in the same address family (IPv4/IPv6).

lag_awsLogicalDeviceId :: Lens' Lag (Maybe Text)

The Direct Connect endpoint that terminates the logical connection. This device might be different than the device that terminates the physical connection.

lag_allowsHostedConnections :: Lens' Lag (Maybe Bool)

Indicates whether the LAG can host other connections.

lag_encryptionMode :: Lens' Lag (Maybe Text)

The LAG MAC Security (MACsec) encryption mode.

The valid values are no_encrypt, should_encrypt, and must_encrypt.

lag_numberOfConnections :: Lens' Lag (Maybe Int)

The number of physical dedicated connections bundled by the LAG, up to a maximum of 10.

lag_jumboFrameCapable :: Lens' Lag (Maybe Bool)

Indicates whether jumbo frames (9001 MTU) are supported.

lag_lagState :: Lens' Lag (Maybe LagState)

The state of the LAG. The following are the possible values:

requested: The initial state of a LAG. The LAG stays in the requested state until the Letter of Authorization (LOA) is available.

pending: The LAG has been approved and is being initialized.

available: The network link is established and the LAG is ready for use.

down: The network link is down.

deleting: The LAG is being deleted.

deleted: The LAG is deleted.

unknown: The state of the LAG is not available.

lag_ownerAccount :: Lens' Lag (Maybe Text)

The ID of the account that owns the LAG.

lag_region :: Lens' Lag (Maybe Text)

The Region where the connection is located.

lag_macSecKeys :: Lens' Lag (Maybe [MacSecKey])

The MAC Security (MACsec) security keys associated with the LAG.

lag_providerName :: Lens' Lag (Maybe Text)

The name of the service provider associated with the LAG.

lag_awsDeviceV2 :: Lens' Lag (Maybe Text)

The Direct Connect endpoint that hosts the LAG.

lag_tags :: Lens' Lag (Maybe (NonEmpty Tag))

The tags associated with the LAG.