Network.AWS.DirectConnect.DescribeCustomerMetadata

Description

Get and view a list of customer agreements, along with their signed status and whether the customer is an NNIPartner, NNIPartnerV2, or a nonPartner.

Creating a Request

data DescribeCustomerMetadata

See: newDescribeCustomerMetadata smart constructor.

Constructors

DescribeCustomerMetadata'

newDescribeCustomerMetadata :: DescribeCustomerMetadata

Create a value of DescribeCustomerMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

Destructuring the Response

data DescribeCustomerMetadataResponse

See: newDescribeCustomerMetadataResponse smart constructor.

Constructors

DescribeCustomerMetadataResponse'

Fields

nniPartnerType :: Maybe NniPartnerType

The type of network-to-network interface (NNI) partner. The partner type will be one of the following:

V1: This partner can only allocate 50Mbps, 100Mbps, 200Mbps, 300Mbps, 400Mbps, or 500Mbps subgigabit connections.

V2: This partner can only allocate 1GB, 2GB, 5GB, or 10GB hosted connections.

nonPartner: The customer is not a partner.

agreements :: Maybe [CustomerAgreement]

The list of customer agreements.

httpStatus :: Int

The response's http status code.

newDescribeCustomerMetadataResponse

Arguments

:: Int
$sel:httpStatus:DescribeCustomerMetadataResponse'

-> DescribeCustomerMetadataResponse

Create a value of DescribeCustomerMetadataResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nniPartnerType:DescribeCustomerMetadataResponse', describeCustomerMetadataResponse_nniPartnerType - The type of network-to-network interface (NNI) partner. The partner type will be one of the following:

V1: This partner can only allocate 50Mbps, 100Mbps, 200Mbps, 300Mbps, 400Mbps, or 500Mbps subgigabit connections.

V2: This partner can only allocate 1GB, 2GB, 5GB, or 10GB hosted connections.

nonPartner: The customer is not a partner.

$sel:agreements:DescribeCustomerMetadataResponse', describeCustomerMetadataResponse_agreements - The list of customer agreements.

$sel:httpStatus:DescribeCustomerMetadataResponse', describeCustomerMetadataResponse_httpStatus - The response's http status code.

Response Lenses

describeCustomerMetadataResponse_nniPartnerType :: Lens' DescribeCustomerMetadataResponse (Maybe NniPartnerType)

The type of network-to-network interface (NNI) partner. The partner type will be one of the following:

V1: This partner can only allocate 50Mbps, 100Mbps, 200Mbps, 300Mbps, 400Mbps, or 500Mbps subgigabit connections.

V2: This partner can only allocate 1GB, 2GB, 5GB, or 10GB hosted connections.

nonPartner: The customer is not a partner.

describeCustomerMetadataResponse_agreements :: Lens' DescribeCustomerMetadataResponse (Maybe [CustomerAgreement])

The list of customer agreements.

describeCustomerMetadataResponse_httpStatus :: Lens' DescribeCustomerMetadataResponse Int

The response's http status code.