Network.AWS.DirectConnect.DescribeDirectConnectGatewayAssociationProposals

Description

Describes one or more association proposals for connection between a virtual private gateway or transit gateway and a Direct Connect gateway.

Creating a Request

data DescribeDirectConnectGatewayAssociationProposals

See: newDescribeDirectConnectGatewayAssociationProposals smart constructor.

Constructors

DescribeDirectConnectGatewayAssociationProposals'

Fields

associatedGatewayId :: Maybe Text

The ID of the associated gateway.

directConnectGatewayId :: Maybe Text

The ID of the Direct Connect gateway.

proposalId :: Maybe Text

The ID of the proposal.

nextToken :: Maybe Text

The token for the next page of results.

maxResults :: Maybe Int

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

If MaxResults is given a value larger than 100, only 100 results are returned.

newDescribeDirectConnectGatewayAssociationProposals :: DescribeDirectConnectGatewayAssociationProposals

Create a value of DescribeDirectConnectGatewayAssociationProposals with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:associatedGatewayId:DescribeDirectConnectGatewayAssociationProposals', describeDirectConnectGatewayAssociationProposals_associatedGatewayId - The ID of the associated gateway.

$sel:directConnectGatewayId:DescribeDirectConnectGatewayAssociationProposals', describeDirectConnectGatewayAssociationProposals_directConnectGatewayId - The ID of the Direct Connect gateway.

$sel:proposalId:DescribeDirectConnectGatewayAssociationProposals', describeDirectConnectGatewayAssociationProposals_proposalId - The ID of the proposal.

$sel:nextToken:DescribeDirectConnectGatewayAssociationProposals', describeDirectConnectGatewayAssociationProposals_nextToken - The token for the next page of results.

$sel:maxResults:DescribeDirectConnectGatewayAssociationProposals', describeDirectConnectGatewayAssociationProposals_maxResults - The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

If MaxResults is given a value larger than 100, only 100 results are returned.

Request Lenses

describeDirectConnectGatewayAssociationProposals_associatedGatewayId :: Lens' DescribeDirectConnectGatewayAssociationProposals (Maybe Text)

The ID of the associated gateway.

describeDirectConnectGatewayAssociationProposals_directConnectGatewayId :: Lens' DescribeDirectConnectGatewayAssociationProposals (Maybe Text)

The ID of the Direct Connect gateway.

describeDirectConnectGatewayAssociationProposals_proposalId :: Lens' DescribeDirectConnectGatewayAssociationProposals (Maybe Text)

The ID of the proposal.

describeDirectConnectGatewayAssociationProposals_nextToken :: Lens' DescribeDirectConnectGatewayAssociationProposals (Maybe Text)

The token for the next page of results.

describeDirectConnectGatewayAssociationProposals_maxResults :: Lens' DescribeDirectConnectGatewayAssociationProposals (Maybe Int)

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

If MaxResults is given a value larger than 100, only 100 results are returned.

Destructuring the Response

data DescribeDirectConnectGatewayAssociationProposalsResponse

See: newDescribeDirectConnectGatewayAssociationProposalsResponse smart constructor.

Constructors

DescribeDirectConnectGatewayAssociationProposalsResponse'

Fields

directConnectGatewayAssociationProposals :: Maybe [DirectConnectGatewayAssociationProposal]

Describes the Direct Connect gateway association proposals.

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

httpStatus :: Int

The response's http status code.

newDescribeDirectConnectGatewayAssociationProposalsResponse

Arguments

:: Int
$sel:httpStatus:DescribeDirectConnectGatewayAssociationProposalsResponse'

-> DescribeDirectConnectGatewayAssociationProposalsResponse

Create a value of DescribeDirectConnectGatewayAssociationProposalsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:directConnectGatewayAssociationProposals:DescribeDirectConnectGatewayAssociationProposalsResponse', describeDirectConnectGatewayAssociationProposalsResponse_directConnectGatewayAssociationProposals - Describes the Direct Connect gateway association proposals.

$sel:nextToken:DescribeDirectConnectGatewayAssociationProposals', describeDirectConnectGatewayAssociationProposalsResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:httpStatus:DescribeDirectConnectGatewayAssociationProposalsResponse', describeDirectConnectGatewayAssociationProposalsResponse_httpStatus - The response's http status code.

Response Lenses

describeDirectConnectGatewayAssociationProposalsResponse_directConnectGatewayAssociationProposals :: Lens' DescribeDirectConnectGatewayAssociationProposalsResponse (Maybe [DirectConnectGatewayAssociationProposal])

Describes the Direct Connect gateway association proposals.

describeDirectConnectGatewayAssociationProposalsResponse_nextToken :: Lens' DescribeDirectConnectGatewayAssociationProposalsResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

describeDirectConnectGatewayAssociationProposalsResponse_httpStatus :: Lens' DescribeDirectConnectGatewayAssociationProposalsResponse Int

The response's http status code.