Network.AWS.DirectConnect.DescribeDirectConnectGatewayAssociations

Description

Lists the associations between your Direct Connect gateways and virtual private gateways and transit gateways. You must specify one of the following:

A Direct Connect gateway

The response contains all virtual private gateways and transit gateways associated with the Direct Connect gateway.

A virtual private gateway

The response contains the Direct Connect gateway.

A transit gateway

The response contains the Direct Connect gateway.

A Direct Connect gateway and a virtual private gateway

The response contains the association between the Direct Connect gateway and virtual private gateway.

A Direct Connect gateway and a transit gateway

The response contains the association between the Direct Connect gateway and transit gateway.

This operation returns paginated results.

Creating a Request

data DescribeDirectConnectGatewayAssociations

See: newDescribeDirectConnectGatewayAssociations smart constructor.

Constructors

DescribeDirectConnectGatewayAssociations'

Fields

virtualGatewayId :: Maybe Text

The ID of the virtual private gateway or transit gateway.

associationId :: Maybe Text

The ID of the Direct Connect gateway association.

associatedGatewayId :: Maybe Text

The ID of the associated gateway.

directConnectGatewayId :: Maybe Text

The ID of the Direct Connect gateway.

nextToken :: Maybe Text

The token provided in the previous call to retrieve the next page.

maxResults :: Maybe Int

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

If MaxResults is given a value larger than 100, only 100 results are returned.

newDescribeDirectConnectGatewayAssociations :: DescribeDirectConnectGatewayAssociations

Create a value of DescribeDirectConnectGatewayAssociations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:virtualGatewayId:DescribeDirectConnectGatewayAssociations', describeDirectConnectGatewayAssociations_virtualGatewayId - The ID of the virtual private gateway or transit gateway.

$sel:associationId:DescribeDirectConnectGatewayAssociations', describeDirectConnectGatewayAssociations_associationId - The ID of the Direct Connect gateway association.

$sel:associatedGatewayId:DescribeDirectConnectGatewayAssociations', describeDirectConnectGatewayAssociations_associatedGatewayId - The ID of the associated gateway.

$sel:directConnectGatewayId:DescribeDirectConnectGatewayAssociations', describeDirectConnectGatewayAssociations_directConnectGatewayId - The ID of the Direct Connect gateway.

$sel:nextToken:DescribeDirectConnectGatewayAssociations', describeDirectConnectGatewayAssociations_nextToken - The token provided in the previous call to retrieve the next page.

$sel:maxResults:DescribeDirectConnectGatewayAssociations', describeDirectConnectGatewayAssociations_maxResults - The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

If MaxResults is given a value larger than 100, only 100 results are returned.

Request Lenses

describeDirectConnectGatewayAssociations_virtualGatewayId :: Lens' DescribeDirectConnectGatewayAssociations (Maybe Text)

The ID of the virtual private gateway or transit gateway.

describeDirectConnectGatewayAssociations_associationId :: Lens' DescribeDirectConnectGatewayAssociations (Maybe Text)

The ID of the Direct Connect gateway association.

describeDirectConnectGatewayAssociations_associatedGatewayId :: Lens' DescribeDirectConnectGatewayAssociations (Maybe Text)

The ID of the associated gateway.

describeDirectConnectGatewayAssociations_directConnectGatewayId :: Lens' DescribeDirectConnectGatewayAssociations (Maybe Text)

The ID of the Direct Connect gateway.

describeDirectConnectGatewayAssociations_nextToken :: Lens' DescribeDirectConnectGatewayAssociations (Maybe Text)

The token provided in the previous call to retrieve the next page.

describeDirectConnectGatewayAssociations_maxResults :: Lens' DescribeDirectConnectGatewayAssociations (Maybe Int)

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

If MaxResults is given a value larger than 100, only 100 results are returned.

Destructuring the Response

data DescribeDirectConnectGatewayAssociationsResponse

See: newDescribeDirectConnectGatewayAssociationsResponse smart constructor.

Constructors

DescribeDirectConnectGatewayAssociationsResponse'

Fields

nextToken :: Maybe Text

The token to retrieve the next page.

directConnectGatewayAssociations :: Maybe [DirectConnectGatewayAssociation]

Information about the associations.

httpStatus :: Int

The response's http status code.

newDescribeDirectConnectGatewayAssociationsResponse

Arguments

:: Int
$sel:httpStatus:DescribeDirectConnectGatewayAssociationsResponse'

-> DescribeDirectConnectGatewayAssociationsResponse

Create a value of DescribeDirectConnectGatewayAssociationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeDirectConnectGatewayAssociations', describeDirectConnectGatewayAssociationsResponse_nextToken - The token to retrieve the next page.

$sel:directConnectGatewayAssociations:DescribeDirectConnectGatewayAssociationsResponse', describeDirectConnectGatewayAssociationsResponse_directConnectGatewayAssociations - Information about the associations.

$sel:httpStatus:DescribeDirectConnectGatewayAssociationsResponse', describeDirectConnectGatewayAssociationsResponse_httpStatus - The response's http status code.

Response Lenses

describeDirectConnectGatewayAssociationsResponse_nextToken :: Lens' DescribeDirectConnectGatewayAssociationsResponse (Maybe Text)

The token to retrieve the next page.

describeDirectConnectGatewayAssociationsResponse_directConnectGatewayAssociations :: Lens' DescribeDirectConnectGatewayAssociationsResponse (Maybe [DirectConnectGatewayAssociation])

Information about the associations.

describeDirectConnectGatewayAssociationsResponse_httpStatus :: Lens' DescribeDirectConnectGatewayAssociationsResponse Int

The response's http status code.