Network.AWS.DirectConnect.DescribeDirectConnectGatewayAttachments

Description

Lists the attachments between your Direct Connect gateways and virtual interfaces. You must specify a Direct Connect gateway, a virtual interface, or both. If you specify a Direct Connect gateway, the response contains all virtual interfaces attached to the Direct Connect gateway. If you specify a virtual interface, the response contains all Direct Connect gateways attached to the virtual interface. If you specify both, the response contains the attachment between the Direct Connect gateway and the virtual interface.

This operation returns paginated results.

Creating a Request

data DescribeDirectConnectGatewayAttachments

See: newDescribeDirectConnectGatewayAttachments smart constructor.

Constructors

DescribeDirectConnectGatewayAttachments'

Fields

nextToken :: Maybe Text

The token provided in the previous call to retrieve the next page.

maxResults :: Maybe Int

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

If MaxResults is given a value larger than 100, only 100 results are returned.

virtualInterfaceId :: Maybe Text

The ID of the virtual interface.

directConnectGatewayId :: Maybe Text

The ID of the Direct Connect gateway.

newDescribeDirectConnectGatewayAttachments :: DescribeDirectConnectGatewayAttachments

Create a value of DescribeDirectConnectGatewayAttachments with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeDirectConnectGatewayAttachments', describeDirectConnectGatewayAttachments_nextToken - The token provided in the previous call to retrieve the next page.

$sel:maxResults:DescribeDirectConnectGatewayAttachments', describeDirectConnectGatewayAttachments_maxResults - The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

If MaxResults is given a value larger than 100, only 100 results are returned.

$sel:virtualInterfaceId:DescribeDirectConnectGatewayAttachments', describeDirectConnectGatewayAttachments_virtualInterfaceId - The ID of the virtual interface.

$sel:directConnectGatewayId:DescribeDirectConnectGatewayAttachments', describeDirectConnectGatewayAttachments_directConnectGatewayId - The ID of the Direct Connect gateway.

Request Lenses

describeDirectConnectGatewayAttachments_nextToken :: Lens' DescribeDirectConnectGatewayAttachments (Maybe Text)

The token provided in the previous call to retrieve the next page.

describeDirectConnectGatewayAttachments_maxResults :: Lens' DescribeDirectConnectGatewayAttachments (Maybe Int)

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

If MaxResults is given a value larger than 100, only 100 results are returned.

describeDirectConnectGatewayAttachments_virtualInterfaceId :: Lens' DescribeDirectConnectGatewayAttachments (Maybe Text)

The ID of the virtual interface.

describeDirectConnectGatewayAttachments_directConnectGatewayId :: Lens' DescribeDirectConnectGatewayAttachments (Maybe Text)

The ID of the Direct Connect gateway.

Destructuring the Response

data DescribeDirectConnectGatewayAttachmentsResponse

See: newDescribeDirectConnectGatewayAttachmentsResponse smart constructor.

Constructors

DescribeDirectConnectGatewayAttachmentsResponse'

Fields

nextToken :: Maybe Text

The token to retrieve the next page.

directConnectGatewayAttachments :: Maybe [DirectConnectGatewayAttachment]

The attachments.

httpStatus :: Int

The response's http status code.

newDescribeDirectConnectGatewayAttachmentsResponse

Arguments

:: Int
$sel:httpStatus:DescribeDirectConnectGatewayAttachmentsResponse'

-> DescribeDirectConnectGatewayAttachmentsResponse

Create a value of DescribeDirectConnectGatewayAttachmentsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeDirectConnectGatewayAttachments', describeDirectConnectGatewayAttachmentsResponse_nextToken - The token to retrieve the next page.

$sel:directConnectGatewayAttachments:DescribeDirectConnectGatewayAttachmentsResponse', describeDirectConnectGatewayAttachmentsResponse_directConnectGatewayAttachments - The attachments.

$sel:httpStatus:DescribeDirectConnectGatewayAttachmentsResponse', describeDirectConnectGatewayAttachmentsResponse_httpStatus - The response's http status code.

Response Lenses

describeDirectConnectGatewayAttachmentsResponse_nextToken :: Lens' DescribeDirectConnectGatewayAttachmentsResponse (Maybe Text)

The token to retrieve the next page.

describeDirectConnectGatewayAttachmentsResponse_directConnectGatewayAttachments :: Lens' DescribeDirectConnectGatewayAttachmentsResponse (Maybe [DirectConnectGatewayAttachment])

The attachments.

describeDirectConnectGatewayAttachmentsResponse_httpStatus :: Lens' DescribeDirectConnectGatewayAttachmentsResponse Int

The response's http status code.