Network.AWS.DirectConnect.DescribeDirectConnectGateways

Description

Lists all your Direct Connect gateways or only the specified Direct Connect gateway. Deleted Direct Connect gateways are not returned.

This operation returns paginated results.

Creating a Request

data DescribeDirectConnectGateways

See: newDescribeDirectConnectGateways smart constructor.

Constructors

DescribeDirectConnectGateways'

Fields

directConnectGatewayId :: Maybe Text

The ID of the Direct Connect gateway.

nextToken :: Maybe Text

The token provided in the previous call to retrieve the next page.

maxResults :: Maybe Int

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

If MaxResults is given a value larger than 100, only 100 results are returned.

newDescribeDirectConnectGateways :: DescribeDirectConnectGateways

Create a value of DescribeDirectConnectGateways with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:directConnectGatewayId:DescribeDirectConnectGateways', describeDirectConnectGateways_directConnectGatewayId - The ID of the Direct Connect gateway.

$sel:nextToken:DescribeDirectConnectGateways', describeDirectConnectGateways_nextToken - The token provided in the previous call to retrieve the next page.

$sel:maxResults:DescribeDirectConnectGateways', describeDirectConnectGateways_maxResults - The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

If MaxResults is given a value larger than 100, only 100 results are returned.

Request Lenses

describeDirectConnectGateways_directConnectGatewayId :: Lens' DescribeDirectConnectGateways (Maybe Text)

The ID of the Direct Connect gateway.

describeDirectConnectGateways_nextToken :: Lens' DescribeDirectConnectGateways (Maybe Text)

The token provided in the previous call to retrieve the next page.

describeDirectConnectGateways_maxResults :: Lens' DescribeDirectConnectGateways (Maybe Int)

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

If MaxResults is given a value larger than 100, only 100 results are returned.

Destructuring the Response

data DescribeDirectConnectGatewaysResponse

See: newDescribeDirectConnectGatewaysResponse smart constructor.

Constructors

DescribeDirectConnectGatewaysResponse'

Fields

directConnectGateways :: Maybe [DirectConnectGateway]

The Direct Connect gateways.

nextToken :: Maybe Text

The token to retrieve the next page.

httpStatus :: Int

The response's http status code.

newDescribeDirectConnectGatewaysResponse

Arguments

:: Int
$sel:httpStatus:DescribeDirectConnectGatewaysResponse'

-> DescribeDirectConnectGatewaysResponse

Create a value of DescribeDirectConnectGatewaysResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:directConnectGateways:DescribeDirectConnectGatewaysResponse', describeDirectConnectGatewaysResponse_directConnectGateways - The Direct Connect gateways.

$sel:nextToken:DescribeDirectConnectGateways', describeDirectConnectGatewaysResponse_nextToken - The token to retrieve the next page.

$sel:httpStatus:DescribeDirectConnectGatewaysResponse', describeDirectConnectGatewaysResponse_httpStatus - The response's http status code.

Response Lenses

describeDirectConnectGatewaysResponse_directConnectGateways :: Lens' DescribeDirectConnectGatewaysResponse (Maybe [DirectConnectGateway])

The Direct Connect gateways.

describeDirectConnectGatewaysResponse_nextToken :: Lens' DescribeDirectConnectGatewaysResponse (Maybe Text)

The token to retrieve the next page.

describeDirectConnectGatewaysResponse_httpStatus :: Lens' DescribeDirectConnectGatewaysResponse Int

The response's http status code.