Network.AWS.DirectConnect.DescribeRouterConfiguration

Description

Details about the router.

Creating a Request

data DescribeRouterConfiguration

Provides the details about a virtual interface's router.

See: newDescribeRouterConfiguration smart constructor.

Constructors

DescribeRouterConfiguration'

Fields

routerTypeIdentifier :: Maybe Text

Identifies the router by a combination of vendor, platform, and software version. For example, CiscoSystemsInc-2900SeriesRouters-IOS124.

virtualInterfaceId :: Text

The ID of the virtual interface.

newDescribeRouterConfiguration

Arguments

:: Text
$sel:virtualInterfaceId:DescribeRouterConfiguration'

-> DescribeRouterConfiguration

Create a value of DescribeRouterConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:routerTypeIdentifier:DescribeRouterConfiguration', describeRouterConfiguration_routerTypeIdentifier - Identifies the router by a combination of vendor, platform, and software version. For example, CiscoSystemsInc-2900SeriesRouters-IOS124.

$sel:virtualInterfaceId:DescribeRouterConfiguration', describeRouterConfiguration_virtualInterfaceId - The ID of the virtual interface.

Request Lenses

describeRouterConfiguration_routerTypeIdentifier :: Lens' DescribeRouterConfiguration (Maybe Text)

Identifies the router by a combination of vendor, platform, and software version. For example, CiscoSystemsInc-2900SeriesRouters-IOS124.

describeRouterConfiguration_virtualInterfaceId :: Lens' DescribeRouterConfiguration Text

The ID of the virtual interface.

Destructuring the Response

data DescribeRouterConfigurationResponse

See: newDescribeRouterConfigurationResponse smart constructor.

Constructors

DescribeRouterConfigurationResponse'

Fields

router :: Maybe RouterType

The details about the router.

customerRouterConfig :: Maybe Text

The customer router configuration.

virtualInterfaceName :: Maybe Text

The name of the virtual interface assigned by the customer network.

virtualInterfaceId :: Maybe Text

The ID assigned to the virtual interface.

httpStatus :: Int

The response's http status code.

newDescribeRouterConfigurationResponse

Arguments

:: Int
$sel:httpStatus:DescribeRouterConfigurationResponse'

-> DescribeRouterConfigurationResponse

Create a value of DescribeRouterConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:router:DescribeRouterConfigurationResponse', describeRouterConfigurationResponse_router - The details about the router.

$sel:customerRouterConfig:DescribeRouterConfigurationResponse', describeRouterConfigurationResponse_customerRouterConfig - The customer router configuration.

$sel:virtualInterfaceName:DescribeRouterConfigurationResponse', describeRouterConfigurationResponse_virtualInterfaceName - The name of the virtual interface assigned by the customer network.

$sel:virtualInterfaceId:DescribeRouterConfiguration', describeRouterConfigurationResponse_virtualInterfaceId - The ID assigned to the virtual interface.

$sel:httpStatus:DescribeRouterConfigurationResponse', describeRouterConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

describeRouterConfigurationResponse_router :: Lens' DescribeRouterConfigurationResponse (Maybe RouterType)

The details about the router.

describeRouterConfigurationResponse_customerRouterConfig :: Lens' DescribeRouterConfigurationResponse (Maybe Text)

The customer router configuration.

describeRouterConfigurationResponse_virtualInterfaceName :: Lens' DescribeRouterConfigurationResponse (Maybe Text)

The name of the virtual interface assigned by the customer network.

describeRouterConfigurationResponse_virtualInterfaceId :: Lens' DescribeRouterConfigurationResponse (Maybe Text)

The ID assigned to the virtual interface.

describeRouterConfigurationResponse_httpStatus :: Lens' DescribeRouterConfigurationResponse Int

The response's http status code.