Network.AWS.DirectConnect.DisassociateConnectionFromLag

Description

Disassociates a connection from a link aggregation group (LAG). The connection is interrupted and re-established as a standalone connection (the connection is not deleted; to delete the connection, use the DeleteConnection request). If the LAG has associated virtual interfaces or hosted connections, they remain associated with the LAG. A disassociated connection owned by an Direct Connect Partner is automatically converted to an interconnect.

If disassociating the connection would cause the LAG to fall below its setting for minimum number of operational connections, the request fails, except when it's the last member of the LAG. If all connections are disassociated, the LAG continues to exist as an empty LAG with no physical connections.

Creating a Request

data DisassociateConnectionFromLag

See: newDisassociateConnectionFromLag smart constructor.

Constructors

DisassociateConnectionFromLag'

Fields

connectionId :: Text

The ID of the connection.

lagId :: Text

The ID of the LAG.

newDisassociateConnectionFromLag

Arguments

:: Text
$sel:connectionId:DisassociateConnectionFromLag'

-> Text
$sel:lagId:DisassociateConnectionFromLag'

-> DisassociateConnectionFromLag

Create a value of DisassociateConnectionFromLag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:connectionId:DisassociateConnectionFromLag', disassociateConnectionFromLag_connectionId - The ID of the connection.

$sel:lagId:DisassociateConnectionFromLag', disassociateConnectionFromLag_lagId - The ID of the LAG.

Request Lenses

disassociateConnectionFromLag_connectionId :: Lens' DisassociateConnectionFromLag Text

The ID of the connection.

disassociateConnectionFromLag_lagId :: Lens' DisassociateConnectionFromLag Text

The ID of the LAG.

Destructuring the Response

data Connection

Information about an Direct Connect connection.

See: newConnection smart constructor.

Constructors

Connection'

Fields

lagId :: Maybe Text

The ID of the LAG.

macSecCapable :: Maybe Bool

Indicates whether the connection supports MAC Security (MACsec).

portEncryptionStatus :: Maybe Text

The MAC Security (MACsec) port link status of the connection.

The valid values are Encryption Up, which means that there is an active Connection Key Name, or Encryption Down.

vlan :: Maybe Int

The ID of the VLAN.

location :: Maybe Text

The location of the connection.

awsDevice :: Maybe Text

The Direct Connect endpoint on which the physical connection terminates.

hasLogicalRedundancy :: Maybe HasLogicalRedundancy

Indicates whether the connection supports a secondary BGP peer in the same address family (IPv4/IPv6).

connectionId :: Maybe Text

The ID of the connection.

awsLogicalDeviceId :: Maybe Text

The Direct Connect endpoint that terminates the logical connection. This device might be different than the device that terminates the physical connection.

loaIssueTime :: Maybe POSIX

The time of the most recent call to DescribeLoa for this connection.

partnerName :: Maybe Text

The name of the Direct Connect service provider associated with the connection.

connectionName :: Maybe Text

The name of the connection.

encryptionMode :: Maybe Text

The MAC Security (MACsec) connection encryption mode.

The valid values are no_encrypt, should_encrypt, and must_encrypt.

bandwidth :: Maybe Text

The bandwidth of the connection.

jumboFrameCapable :: Maybe Bool

Indicates whether jumbo frames (9001 MTU) are supported.

ownerAccount :: Maybe Text

The ID of the Amazon Web Services account that owns the connection.

region :: Maybe Text

The Amazon Web Services Region where the connection is located.

macSecKeys :: Maybe [MacSecKey]

The MAC Security (MACsec) security keys associated with the connection.

providerName :: Maybe Text

The name of the service provider associated with the connection.

awsDeviceV2 :: Maybe Text

The Direct Connect endpoint that terminates the physical connection.

connectionState :: Maybe ConnectionState

The state of the connection. The following are the possible values:

ordering: The initial state of a hosted connection provisioned on an interconnect. The connection stays in the ordering state until the owner of the hosted connection confirms or declines the connection order.

requested: The initial state of a standard connection. The connection stays in the requested state until the Letter of Authorization (LOA) is sent to the customer.

pending: The connection has been approved and is being initialized.

available: The network link is up and the connection is ready for use.

down: The network link is down.

deleting: The connection is being deleted.

deleted: The connection has been deleted.

rejected: A hosted connection in the ordering state enters the rejected state if it is deleted by the customer.

unknown: The state of the connection is not available.

tags :: Maybe (NonEmpty Tag)

The tags associated with the connection.

newConnection :: Connection

Create a value of Connection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lagId:Connection', connection_lagId - The ID of the LAG.

$sel:macSecCapable:Connection', connection_macSecCapable - Indicates whether the connection supports MAC Security (MACsec).

$sel:portEncryptionStatus:Connection', connection_portEncryptionStatus - The MAC Security (MACsec) port link status of the connection.

The valid values are Encryption Up, which means that there is an active Connection Key Name, or Encryption Down.

$sel:vlan:Connection', connection_vlan - The ID of the VLAN.

$sel:location:Connection', connection_location - The location of the connection.

$sel:awsDevice:Connection', connection_awsDevice - The Direct Connect endpoint on which the physical connection terminates.

$sel:hasLogicalRedundancy:Connection', connection_hasLogicalRedundancy - Indicates whether the connection supports a secondary BGP peer in the same address family (IPv4/IPv6).

$sel:connectionId:Connection', connection_connectionId - The ID of the connection.

$sel:awsLogicalDeviceId:Connection', connection_awsLogicalDeviceId - The Direct Connect endpoint that terminates the logical connection. This device might be different than the device that terminates the physical connection.

$sel:loaIssueTime:Connection', connection_loaIssueTime - The time of the most recent call to DescribeLoa for this connection.

$sel:partnerName:Connection', connection_partnerName - The name of the Direct Connect service provider associated with the connection.

$sel:connectionName:Connection', connection_connectionName - The name of the connection.

$sel:encryptionMode:Connection', connection_encryptionMode - The MAC Security (MACsec) connection encryption mode.

The valid values are no_encrypt, should_encrypt, and must_encrypt.

$sel:bandwidth:Connection', connection_bandwidth - The bandwidth of the connection.

$sel:jumboFrameCapable:Connection', connection_jumboFrameCapable - Indicates whether jumbo frames (9001 MTU) are supported.

$sel:ownerAccount:Connection', connection_ownerAccount - The ID of the Amazon Web Services account that owns the connection.

$sel:region:Connection', connection_region - The Amazon Web Services Region where the connection is located.

$sel:macSecKeys:Connection', connection_macSecKeys - The MAC Security (MACsec) security keys associated with the connection.

$sel:providerName:Connection', connection_providerName - The name of the service provider associated with the connection.

$sel:awsDeviceV2:Connection', connection_awsDeviceV2 - The Direct Connect endpoint that terminates the physical connection.

$sel:connectionState:Connection', connection_connectionState - The state of the connection. The following are the possible values:

ordering: The initial state of a hosted connection provisioned on an interconnect. The connection stays in the ordering state until the owner of the hosted connection confirms or declines the connection order.

requested: The initial state of a standard connection. The connection stays in the requested state until the Letter of Authorization (LOA) is sent to the customer.

pending: The connection has been approved and is being initialized.

available: The network link is up and the connection is ready for use.

down: The network link is down.

deleting: The connection is being deleted.

deleted: The connection has been deleted.

rejected: A hosted connection in the ordering state enters the rejected state if it is deleted by the customer.

unknown: The state of the connection is not available.

$sel:tags:Connection', connection_tags - The tags associated with the connection.

Response Lenses

connection_lagId :: Lens' Connection (Maybe Text)

The ID of the LAG.

connection_macSecCapable :: Lens' Connection (Maybe Bool)

Indicates whether the connection supports MAC Security (MACsec).

connection_portEncryptionStatus :: Lens' Connection (Maybe Text)

The MAC Security (MACsec) port link status of the connection.

The valid values are Encryption Up, which means that there is an active Connection Key Name, or Encryption Down.

connection_vlan :: Lens' Connection (Maybe Int)

The ID of the VLAN.

connection_location :: Lens' Connection (Maybe Text)

The location of the connection.

connection_awsDevice :: Lens' Connection (Maybe Text)

The Direct Connect endpoint on which the physical connection terminates.

connection_hasLogicalRedundancy :: Lens' Connection (Maybe HasLogicalRedundancy)

Indicates whether the connection supports a secondary BGP peer in the same address family (IPv4/IPv6).

connection_connectionId :: Lens' Connection (Maybe Text)

The ID of the connection.

connection_awsLogicalDeviceId :: Lens' Connection (Maybe Text)

The Direct Connect endpoint that terminates the logical connection. This device might be different than the device that terminates the physical connection.

connection_loaIssueTime :: Lens' Connection (Maybe UTCTime)

The time of the most recent call to DescribeLoa for this connection.

connection_partnerName :: Lens' Connection (Maybe Text)

The name of the Direct Connect service provider associated with the connection.

connection_connectionName :: Lens' Connection (Maybe Text)

The name of the connection.

connection_encryptionMode :: Lens' Connection (Maybe Text)

The MAC Security (MACsec) connection encryption mode.

The valid values are no_encrypt, should_encrypt, and must_encrypt.

connection_bandwidth :: Lens' Connection (Maybe Text)

The bandwidth of the connection.

connection_jumboFrameCapable :: Lens' Connection (Maybe Bool)

Indicates whether jumbo frames (9001 MTU) are supported.

connection_ownerAccount :: Lens' Connection (Maybe Text)

The ID of the Amazon Web Services account that owns the connection.

connection_region :: Lens' Connection (Maybe Text)

The Amazon Web Services Region where the connection is located.

connection_macSecKeys :: Lens' Connection (Maybe [MacSecKey])

The MAC Security (MACsec) security keys associated with the connection.

connection_providerName :: Lens' Connection (Maybe Text)

The name of the service provider associated with the connection.

connection_awsDeviceV2 :: Lens' Connection (Maybe Text)

The Direct Connect endpoint that terminates the physical connection.

connection_connectionState :: Lens' Connection (Maybe ConnectionState)

The state of the connection. The following are the possible values:

ordering: The initial state of a hosted connection provisioned on an interconnect. The connection stays in the ordering state until the owner of the hosted connection confirms or declines the connection order.

requested: The initial state of a standard connection. The connection stays in the requested state until the Letter of Authorization (LOA) is sent to the customer.

pending: The connection has been approved and is being initialized.

available: The network link is up and the connection is ready for use.

down: The network link is down.

deleting: The connection is being deleted.

deleted: The connection has been deleted.

rejected: A hosted connection in the ordering state enters the rejected state if it is deleted by the customer.

unknown: The state of the connection is not available.

connection_tags :: Lens' Connection (Maybe (NonEmpty Tag))

The tags associated with the connection.