Network.AWS.DirectConnect.ListVirtualInterfaceTestHistory

Description

Lists the virtual interface failover test history.

Creating a Request

data ListVirtualInterfaceTestHistory

See: newListVirtualInterfaceTestHistory smart constructor.

Constructors

ListVirtualInterfaceTestHistory'

Fields

bgpPeers :: Maybe [Text]

The BGP peers that were placed in the DOWN state during the virtual interface failover test.

nextToken :: Maybe Text

The token for the next page of results.

status :: Maybe Text

The status of the virtual interface failover test.

maxResults :: Maybe Int

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

If MaxResults is given a value larger than 100, only 100 results are returned.

testId :: Maybe Text

The ID of the virtual interface failover test.

virtualInterfaceId :: Maybe Text

The ID of the virtual interface that was tested.

newListVirtualInterfaceTestHistory :: ListVirtualInterfaceTestHistory

Create a value of ListVirtualInterfaceTestHistory with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bgpPeers:ListVirtualInterfaceTestHistory', listVirtualInterfaceTestHistory_bgpPeers - The BGP peers that were placed in the DOWN state during the virtual interface failover test.

$sel:nextToken:ListVirtualInterfaceTestHistory', listVirtualInterfaceTestHistory_nextToken - The token for the next page of results.

$sel:status:ListVirtualInterfaceTestHistory', listVirtualInterfaceTestHistory_status - The status of the virtual interface failover test.

$sel:maxResults:ListVirtualInterfaceTestHistory', listVirtualInterfaceTestHistory_maxResults - The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

If MaxResults is given a value larger than 100, only 100 results are returned.

$sel:testId:ListVirtualInterfaceTestHistory', listVirtualInterfaceTestHistory_testId - The ID of the virtual interface failover test.

$sel:virtualInterfaceId:ListVirtualInterfaceTestHistory', listVirtualInterfaceTestHistory_virtualInterfaceId - The ID of the virtual interface that was tested.

Request Lenses

listVirtualInterfaceTestHistory_bgpPeers :: Lens' ListVirtualInterfaceTestHistory (Maybe [Text])

The BGP peers that were placed in the DOWN state during the virtual interface failover test.

listVirtualInterfaceTestHistory_nextToken :: Lens' ListVirtualInterfaceTestHistory (Maybe Text)

The token for the next page of results.

listVirtualInterfaceTestHistory_status :: Lens' ListVirtualInterfaceTestHistory (Maybe Text)

The status of the virtual interface failover test.

listVirtualInterfaceTestHistory_maxResults :: Lens' ListVirtualInterfaceTestHistory (Maybe Int)

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

If MaxResults is given a value larger than 100, only 100 results are returned.

listVirtualInterfaceTestHistory_testId :: Lens' ListVirtualInterfaceTestHistory (Maybe Text)

The ID of the virtual interface failover test.

listVirtualInterfaceTestHistory_virtualInterfaceId :: Lens' ListVirtualInterfaceTestHistory (Maybe Text)

The ID of the virtual interface that was tested.

Destructuring the Response

data ListVirtualInterfaceTestHistoryResponse

See: newListVirtualInterfaceTestHistoryResponse smart constructor.

Constructors

ListVirtualInterfaceTestHistoryResponse'

Fields

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

virtualInterfaceTestHistory :: Maybe [VirtualInterfaceTestHistory]

The ID of the tested virtual interface.

httpStatus :: Int

The response's http status code.

newListVirtualInterfaceTestHistoryResponse

Arguments

:: Int
$sel:httpStatus:ListVirtualInterfaceTestHistoryResponse'

-> ListVirtualInterfaceTestHistoryResponse

Create a value of ListVirtualInterfaceTestHistoryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListVirtualInterfaceTestHistory', listVirtualInterfaceTestHistoryResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:virtualInterfaceTestHistory:ListVirtualInterfaceTestHistoryResponse', listVirtualInterfaceTestHistoryResponse_virtualInterfaceTestHistory - The ID of the tested virtual interface.

$sel:httpStatus:ListVirtualInterfaceTestHistoryResponse', listVirtualInterfaceTestHistoryResponse_httpStatus - The response's http status code.

Response Lenses

listVirtualInterfaceTestHistoryResponse_nextToken :: Lens' ListVirtualInterfaceTestHistoryResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

listVirtualInterfaceTestHistoryResponse_virtualInterfaceTestHistory :: Lens' ListVirtualInterfaceTestHistoryResponse (Maybe [VirtualInterfaceTestHistory])

The ID of the tested virtual interface.

listVirtualInterfaceTestHistoryResponse_httpStatus :: Lens' ListVirtualInterfaceTestHistoryResponse Int

The response's http status code.