Network.AWS.DirectConnect.Types.DirectConnectGateway

Description

Documentation

data DirectConnectGateway

Information about a Direct Connect gateway, which enables you to connect virtual interfaces and virtual private gateway or transit gateways.

See: newDirectConnectGateway smart constructor.

Constructors

DirectConnectGateway'

Fields

directConnectGatewayState :: Maybe DirectConnectGatewayState

The state of the Direct Connect gateway. The following are the possible values:

pending: The initial state after calling CreateDirectConnectGateway.

available: The Direct Connect gateway is ready for use.

deleting: The initial state after calling DeleteDirectConnectGateway.

deleted: The Direct Connect gateway is deleted and cannot pass traffic.

stateChangeError :: Maybe Text

The error message if the state of an object failed to advance.

directConnectGatewayName :: Maybe Text

The name of the Direct Connect gateway.

amazonSideAsn :: Maybe Integer

The autonomous system number (ASN) for the Amazon side of the connection.

directConnectGatewayId :: Maybe Text

The ID of the Direct Connect gateway.

ownerAccount :: Maybe Text

The ID of the account that owns the Direct Connect gateway.

newDirectConnectGateway :: DirectConnectGateway

Create a value of DirectConnectGateway with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:directConnectGatewayState:DirectConnectGateway', directConnectGateway_directConnectGatewayState - The state of the Direct Connect gateway. The following are the possible values:

pending: The initial state after calling CreateDirectConnectGateway.

available: The Direct Connect gateway is ready for use.

deleting: The initial state after calling DeleteDirectConnectGateway.

deleted: The Direct Connect gateway is deleted and cannot pass traffic.

$sel:stateChangeError:DirectConnectGateway', directConnectGateway_stateChangeError - The error message if the state of an object failed to advance.

$sel:directConnectGatewayName:DirectConnectGateway', directConnectGateway_directConnectGatewayName - The name of the Direct Connect gateway.

$sel:amazonSideAsn:DirectConnectGateway', directConnectGateway_amazonSideAsn - The autonomous system number (ASN) for the Amazon side of the connection.

$sel:directConnectGatewayId:DirectConnectGateway', directConnectGateway_directConnectGatewayId - The ID of the Direct Connect gateway.

$sel:ownerAccount:DirectConnectGateway', directConnectGateway_ownerAccount - The ID of the account that owns the Direct Connect gateway.

directConnectGateway_directConnectGatewayState :: Lens' DirectConnectGateway (Maybe DirectConnectGatewayState)

The state of the Direct Connect gateway. The following are the possible values:

pending: The initial state after calling CreateDirectConnectGateway.

available: The Direct Connect gateway is ready for use.

deleting: The initial state after calling DeleteDirectConnectGateway.

deleted: The Direct Connect gateway is deleted and cannot pass traffic.

directConnectGateway_stateChangeError :: Lens' DirectConnectGateway (Maybe Text)

The error message if the state of an object failed to advance.

directConnectGateway_directConnectGatewayName :: Lens' DirectConnectGateway (Maybe Text)

The name of the Direct Connect gateway.

directConnectGateway_amazonSideAsn :: Lens' DirectConnectGateway (Maybe Integer)

The autonomous system number (ASN) for the Amazon side of the connection.

directConnectGateway_directConnectGatewayId :: Lens' DirectConnectGateway (Maybe Text)

The ID of the Direct Connect gateway.

directConnectGateway_ownerAccount :: Lens' DirectConnectGateway (Maybe Text)

The ID of the account that owns the Direct Connect gateway.