Network.AWS.DirectConnect.Types.DirectConnectGatewayAssociation

Description

Documentation

data DirectConnectGatewayAssociation

Information about an association between a Direct Connect gateway and a virtual private gateway or transit gateway.

See: newDirectConnectGatewayAssociation smart constructor.

Constructors

DirectConnectGatewayAssociation'

Fields

virtualGatewayId :: Maybe Text

The ID of the virtual private gateway. Applies only to private virtual interfaces.

associationId :: Maybe Text

The ID of the Direct Connect gateway association.

directConnectGatewayId :: Maybe Text

The ID of the Direct Connect gateway.

virtualGatewayOwnerAccount :: Maybe Text

The ID of the account that owns the virtual private gateway.

stateChangeError :: Maybe Text

The error message if the state of an object failed to advance.

virtualGatewayRegion :: Maybe Text

The Region where the virtual private gateway is located.

associatedGateway :: Maybe AssociatedGateway

Information about the associated gateway.

directConnectGatewayOwnerAccount :: Maybe Text

The ID of the account that owns the associated gateway.

allowedPrefixesToDirectConnectGateway :: Maybe [RouteFilterPrefix]

The Amazon VPC prefixes to advertise to the Direct Connect gateway.

associationState :: Maybe DirectConnectGatewayAssociationState

The state of the association. The following are the possible values:

associating: The initial state after calling CreateDirectConnectGatewayAssociation.

associated: The Direct Connect gateway and virtual private gateway or transit gateway are successfully associated and ready to pass traffic.

disassociating: The initial state after calling DeleteDirectConnectGatewayAssociation.

disassociated: The virtual private gateway or transit gateway is disassociated from the Direct Connect gateway. Traffic flow between the Direct Connect gateway and virtual private gateway or transit gateway is stopped.

newDirectConnectGatewayAssociation :: DirectConnectGatewayAssociation

Create a value of DirectConnectGatewayAssociation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:virtualGatewayId:DirectConnectGatewayAssociation', directConnectGatewayAssociation_virtualGatewayId - The ID of the virtual private gateway. Applies only to private virtual interfaces.

$sel:associationId:DirectConnectGatewayAssociation', directConnectGatewayAssociation_associationId - The ID of the Direct Connect gateway association.

$sel:directConnectGatewayId:DirectConnectGatewayAssociation', directConnectGatewayAssociation_directConnectGatewayId - The ID of the Direct Connect gateway.

$sel:virtualGatewayOwnerAccount:DirectConnectGatewayAssociation', directConnectGatewayAssociation_virtualGatewayOwnerAccount - The ID of the account that owns the virtual private gateway.

$sel:stateChangeError:DirectConnectGatewayAssociation', directConnectGatewayAssociation_stateChangeError - The error message if the state of an object failed to advance.

$sel:virtualGatewayRegion:DirectConnectGatewayAssociation', directConnectGatewayAssociation_virtualGatewayRegion - The Region where the virtual private gateway is located.

$sel:associatedGateway:DirectConnectGatewayAssociation', directConnectGatewayAssociation_associatedGateway - Information about the associated gateway.

$sel:directConnectGatewayOwnerAccount:DirectConnectGatewayAssociation', directConnectGatewayAssociation_directConnectGatewayOwnerAccount - The ID of the account that owns the associated gateway.

$sel:allowedPrefixesToDirectConnectGateway:DirectConnectGatewayAssociation', directConnectGatewayAssociation_allowedPrefixesToDirectConnectGateway - The Amazon VPC prefixes to advertise to the Direct Connect gateway.

$sel:associationState:DirectConnectGatewayAssociation', directConnectGatewayAssociation_associationState - The state of the association. The following are the possible values:

associating: The initial state after calling CreateDirectConnectGatewayAssociation.

associated: The Direct Connect gateway and virtual private gateway or transit gateway are successfully associated and ready to pass traffic.

disassociating: The initial state after calling DeleteDirectConnectGatewayAssociation.

disassociated: The virtual private gateway or transit gateway is disassociated from the Direct Connect gateway. Traffic flow between the Direct Connect gateway and virtual private gateway or transit gateway is stopped.

directConnectGatewayAssociation_virtualGatewayId :: Lens' DirectConnectGatewayAssociation (Maybe Text)

The ID of the virtual private gateway. Applies only to private virtual interfaces.

directConnectGatewayAssociation_associationId :: Lens' DirectConnectGatewayAssociation (Maybe Text)

The ID of the Direct Connect gateway association.

directConnectGatewayAssociation_directConnectGatewayId :: Lens' DirectConnectGatewayAssociation (Maybe Text)

The ID of the Direct Connect gateway.

directConnectGatewayAssociation_virtualGatewayOwnerAccount :: Lens' DirectConnectGatewayAssociation (Maybe Text)

The ID of the account that owns the virtual private gateway.

directConnectGatewayAssociation_stateChangeError :: Lens' DirectConnectGatewayAssociation (Maybe Text)

The error message if the state of an object failed to advance.

directConnectGatewayAssociation_virtualGatewayRegion :: Lens' DirectConnectGatewayAssociation (Maybe Text)

The Region where the virtual private gateway is located.

directConnectGatewayAssociation_associatedGateway :: Lens' DirectConnectGatewayAssociation (Maybe AssociatedGateway)

Information about the associated gateway.

directConnectGatewayAssociation_directConnectGatewayOwnerAccount :: Lens' DirectConnectGatewayAssociation (Maybe Text)

The ID of the account that owns the associated gateway.

directConnectGatewayAssociation_allowedPrefixesToDirectConnectGateway :: Lens' DirectConnectGatewayAssociation (Maybe [RouteFilterPrefix])

The Amazon VPC prefixes to advertise to the Direct Connect gateway.

directConnectGatewayAssociation_associationState :: Lens' DirectConnectGatewayAssociation (Maybe DirectConnectGatewayAssociationState)

The state of the association. The following are the possible values:

associating: The initial state after calling CreateDirectConnectGatewayAssociation.

associated: The Direct Connect gateway and virtual private gateway or transit gateway are successfully associated and ready to pass traffic.

disassociating: The initial state after calling DeleteDirectConnectGatewayAssociation.

disassociated: The virtual private gateway or transit gateway is disassociated from the Direct Connect gateway. Traffic flow between the Direct Connect gateway and virtual private gateway or transit gateway is stopped.