Network.AWS.DirectConnect.Types.DirectConnectGatewayAttachment

Description

Documentation

data DirectConnectGatewayAttachment

Information about an attachment between a Direct Connect gateway and a virtual interface.

See: newDirectConnectGatewayAttachment smart constructor.

Constructors

DirectConnectGatewayAttachment'

Fields

directConnectGatewayId :: Maybe Text

The ID of the Direct Connect gateway.

attachmentState :: Maybe DirectConnectGatewayAttachmentState

The state of the attachment. The following are the possible values:

attaching: The initial state after a virtual interface is created using the Direct Connect gateway.

attached: The Direct Connect gateway and virtual interface are attached and ready to pass traffic.

detaching: The initial state after calling DeleteVirtualInterface.

detached: The virtual interface is detached from the Direct Connect gateway. Traffic flow between the Direct Connect gateway and virtual interface is stopped.

stateChangeError :: Maybe Text

The error message if the state of an object failed to advance.

virtualInterfaceRegion :: Maybe Text

The Region where the virtual interface is located.

virtualInterfaceOwnerAccount :: Maybe Text

The ID of the account that owns the virtual interface.

virtualInterfaceId :: Maybe Text

The ID of the virtual interface.

attachmentType :: Maybe DirectConnectGatewayAttachmentType

The type of attachment.

newDirectConnectGatewayAttachment :: DirectConnectGatewayAttachment

Create a value of DirectConnectGatewayAttachment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:directConnectGatewayId:DirectConnectGatewayAttachment', directConnectGatewayAttachment_directConnectGatewayId - The ID of the Direct Connect gateway.

$sel:attachmentState:DirectConnectGatewayAttachment', directConnectGatewayAttachment_attachmentState - The state of the attachment. The following are the possible values:

attaching: The initial state after a virtual interface is created using the Direct Connect gateway.

attached: The Direct Connect gateway and virtual interface are attached and ready to pass traffic.

detaching: The initial state after calling DeleteVirtualInterface.

detached: The virtual interface is detached from the Direct Connect gateway. Traffic flow between the Direct Connect gateway and virtual interface is stopped.

$sel:stateChangeError:DirectConnectGatewayAttachment', directConnectGatewayAttachment_stateChangeError - The error message if the state of an object failed to advance.

$sel:virtualInterfaceRegion:DirectConnectGatewayAttachment', directConnectGatewayAttachment_virtualInterfaceRegion - The Region where the virtual interface is located.

$sel:virtualInterfaceOwnerAccount:DirectConnectGatewayAttachment', directConnectGatewayAttachment_virtualInterfaceOwnerAccount - The ID of the account that owns the virtual interface.

$sel:virtualInterfaceId:DirectConnectGatewayAttachment', directConnectGatewayAttachment_virtualInterfaceId - The ID of the virtual interface.

$sel:attachmentType:DirectConnectGatewayAttachment', directConnectGatewayAttachment_attachmentType - The type of attachment.

directConnectGatewayAttachment_directConnectGatewayId :: Lens' DirectConnectGatewayAttachment (Maybe Text)

The ID of the Direct Connect gateway.

directConnectGatewayAttachment_attachmentState :: Lens' DirectConnectGatewayAttachment (Maybe DirectConnectGatewayAttachmentState)

The state of the attachment. The following are the possible values:

attaching: The initial state after a virtual interface is created using the Direct Connect gateway.

attached: The Direct Connect gateway and virtual interface are attached and ready to pass traffic.

detaching: The initial state after calling DeleteVirtualInterface.

detached: The virtual interface is detached from the Direct Connect gateway. Traffic flow between the Direct Connect gateway and virtual interface is stopped.

directConnectGatewayAttachment_stateChangeError :: Lens' DirectConnectGatewayAttachment (Maybe Text)

The error message if the state of an object failed to advance.

directConnectGatewayAttachment_virtualInterfaceRegion :: Lens' DirectConnectGatewayAttachment (Maybe Text)

The Region where the virtual interface is located.

directConnectGatewayAttachment_virtualInterfaceOwnerAccount :: Lens' DirectConnectGatewayAttachment (Maybe Text)

The ID of the account that owns the virtual interface.

directConnectGatewayAttachment_virtualInterfaceId :: Lens' DirectConnectGatewayAttachment (Maybe Text)

The ID of the virtual interface.

directConnectGatewayAttachment_attachmentType :: Lens' DirectConnectGatewayAttachment (Maybe DirectConnectGatewayAttachmentType)

The type of attachment.