Network.AWS.DirectConnect.Types.Interconnect

Description

Documentation

data Interconnect

Information about an interconnect.

See: newInterconnect smart constructor.

Constructors

Interconnect'

Fields

interconnectId :: Maybe Text

The ID of the interconnect.

bandwidth :: Maybe Text

The bandwidth of the connection.

awsDeviceV2 :: Maybe Text

The Direct Connect endpoint that terminates the physical connection.

providerName :: Maybe Text

The name of the service provider associated with the interconnect.

awsLogicalDeviceId :: Maybe Text

The Direct Connect endpoint that terminates the logical connection. This device might be different than the device that terminates the physical connection.

hasLogicalRedundancy :: Maybe HasLogicalRedundancy

Indicates whether the interconnect supports a secondary BGP in the same address family (IPv4/IPv6).

awsDevice :: Maybe Text

The Direct Connect endpoint on which the physical connection terminates.

jumboFrameCapable :: Maybe Bool

Indicates whether jumbo frames (9001 MTU) are supported.

lagId :: Maybe Text

The ID of the LAG.

tags :: Maybe (NonEmpty Tag)

The tags associated with the interconnect.

loaIssueTime :: Maybe POSIX

The time of the most recent call to DescribeLoa for this connection.

region :: Maybe Text

The Region where the connection is located.

interconnectName :: Maybe Text

The name of the interconnect.

interconnectState :: Maybe InterconnectState

The state of the interconnect. The following are the possible values:

requested: The initial state of an interconnect. The interconnect stays in the requested state until the Letter of Authorization (LOA) is sent to the customer.

pending: The interconnect is approved, and is being initialized.

available: The network link is up, and the interconnect is ready for use.

down: The network link is down.

deleting: The interconnect is being deleted.

deleted: The interconnect is deleted.

unknown: The state of the interconnect is not available.

location :: Maybe Text

The location of the connection.

newInterconnect :: Interconnect

Create a value of Interconnect with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:interconnectId:Interconnect', interconnect_interconnectId - The ID of the interconnect.

$sel:bandwidth:Interconnect', interconnect_bandwidth - The bandwidth of the connection.

$sel:awsDeviceV2:Interconnect', interconnect_awsDeviceV2 - The Direct Connect endpoint that terminates the physical connection.

$sel:providerName:Interconnect', interconnect_providerName - The name of the service provider associated with the interconnect.

$sel:awsLogicalDeviceId:Interconnect', interconnect_awsLogicalDeviceId - The Direct Connect endpoint that terminates the logical connection. This device might be different than the device that terminates the physical connection.

$sel:hasLogicalRedundancy:Interconnect', interconnect_hasLogicalRedundancy - Indicates whether the interconnect supports a secondary BGP in the same address family (IPv4/IPv6).

$sel:awsDevice:Interconnect', interconnect_awsDevice - The Direct Connect endpoint on which the physical connection terminates.

$sel:jumboFrameCapable:Interconnect', interconnect_jumboFrameCapable - Indicates whether jumbo frames (9001 MTU) are supported.

$sel:lagId:Interconnect', interconnect_lagId - The ID of the LAG.

$sel:tags:Interconnect', interconnect_tags - The tags associated with the interconnect.

$sel:loaIssueTime:Interconnect', interconnect_loaIssueTime - The time of the most recent call to DescribeLoa for this connection.

$sel:region:Interconnect', interconnect_region - The Region where the connection is located.

$sel:interconnectName:Interconnect', interconnect_interconnectName - The name of the interconnect.

$sel:interconnectState:Interconnect', interconnect_interconnectState - The state of the interconnect. The following are the possible values:

requested: The initial state of an interconnect. The interconnect stays in the requested state until the Letter of Authorization (LOA) is sent to the customer.

pending: The interconnect is approved, and is being initialized.

available: The network link is up, and the interconnect is ready for use.

down: The network link is down.

deleting: The interconnect is being deleted.

deleted: The interconnect is deleted.

unknown: The state of the interconnect is not available.

$sel:location:Interconnect', interconnect_location - The location of the connection.

interconnect_interconnectId :: Lens' Interconnect (Maybe Text)

The ID of the interconnect.

interconnect_bandwidth :: Lens' Interconnect (Maybe Text)

The bandwidth of the connection.

interconnect_awsDeviceV2 :: Lens' Interconnect (Maybe Text)

The Direct Connect endpoint that terminates the physical connection.

interconnect_providerName :: Lens' Interconnect (Maybe Text)

The name of the service provider associated with the interconnect.

interconnect_awsLogicalDeviceId :: Lens' Interconnect (Maybe Text)

The Direct Connect endpoint that terminates the logical connection. This device might be different than the device that terminates the physical connection.

interconnect_hasLogicalRedundancy :: Lens' Interconnect (Maybe HasLogicalRedundancy)

Indicates whether the interconnect supports a secondary BGP in the same address family (IPv4/IPv6).

interconnect_awsDevice :: Lens' Interconnect (Maybe Text)

The Direct Connect endpoint on which the physical connection terminates.

interconnect_jumboFrameCapable :: Lens' Interconnect (Maybe Bool)

Indicates whether jumbo frames (9001 MTU) are supported.

interconnect_lagId :: Lens' Interconnect (Maybe Text)

The ID of the LAG.

interconnect_tags :: Lens' Interconnect (Maybe (NonEmpty Tag))

The tags associated with the interconnect.

interconnect_loaIssueTime :: Lens' Interconnect (Maybe UTCTime)

The time of the most recent call to DescribeLoa for this connection.

interconnect_region :: Lens' Interconnect (Maybe Text)

The Region where the connection is located.

interconnect_interconnectName :: Lens' Interconnect (Maybe Text)

The name of the interconnect.

interconnect_interconnectState :: Lens' Interconnect (Maybe InterconnectState)

The state of the interconnect. The following are the possible values:

requested: The initial state of an interconnect. The interconnect stays in the requested state until the Letter of Authorization (LOA) is sent to the customer.

pending: The interconnect is approved, and is being initialized.

available: The network link is up, and the interconnect is ready for use.

down: The network link is down.

deleting: The interconnect is being deleted.

deleted: The interconnect is deleted.

unknown: The state of the interconnect is not available.

interconnect_location :: Lens' Interconnect (Maybe Text)

The location of the connection.