Network.AWS.DirectConnect.Types.NewPrivateVirtualInterface

Description

Documentation

data NewPrivateVirtualInterface

Information about a private virtual interface.

See: newNewPrivateVirtualInterface smart constructor.

Constructors

NewPrivateVirtualInterface'

Fields

authKey :: Maybe Text

The authentication key for BGP configuration. This string has a minimum length of 6 characters and and a maximun lenth of 80 characters.

virtualGatewayId :: Maybe Text

The ID of the virtual private gateway.

mtu :: Maybe Int

The maximum transmission unit (MTU), in bytes. The supported values are 1500 and 9001. The default value is 1500.

tags :: Maybe (NonEmpty Tag)

The tags associated with the private virtual interface.

directConnectGatewayId :: Maybe Text

The ID of the Direct Connect gateway.

amazonAddress :: Maybe Text

The IP address assigned to the Amazon interface.

addressFamily :: Maybe AddressFamily

The address family for the BGP peer.

customerAddress :: Maybe Text

The IP address assigned to the customer interface.

virtualInterfaceName :: Text

The name of the virtual interface assigned by the customer network. The name has a maximum of 100 characters. The following are valid characters: a-z, 0-9 and a hyphen (-).

vlan :: Int

The ID of the VLAN.

asn :: Int

The autonomous system (AS) number for Border Gateway Protocol (BGP) configuration.

The valid values are 1-2147483647.

newNewPrivateVirtualInterface

Arguments

:: Text
$sel:virtualInterfaceName:NewPrivateVirtualInterface'

-> Int
$sel:vlan:NewPrivateVirtualInterface'

-> Int
$sel:asn:NewPrivateVirtualInterface'

-> NewPrivateVirtualInterface

Create a value of NewPrivateVirtualInterface with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:authKey:NewPrivateVirtualInterface', newPrivateVirtualInterface_authKey - The authentication key for BGP configuration. This string has a minimum length of 6 characters and and a maximun lenth of 80 characters.

$sel:virtualGatewayId:NewPrivateVirtualInterface', newPrivateVirtualInterface_virtualGatewayId - The ID of the virtual private gateway.

$sel:mtu:NewPrivateVirtualInterface', newPrivateVirtualInterface_mtu - The maximum transmission unit (MTU), in bytes. The supported values are 1500 and 9001. The default value is 1500.

$sel:tags:NewPrivateVirtualInterface', newPrivateVirtualInterface_tags - The tags associated with the private virtual interface.

$sel:directConnectGatewayId:NewPrivateVirtualInterface', newPrivateVirtualInterface_directConnectGatewayId - The ID of the Direct Connect gateway.

$sel:amazonAddress:NewPrivateVirtualInterface', newPrivateVirtualInterface_amazonAddress - The IP address assigned to the Amazon interface.

$sel:addressFamily:NewPrivateVirtualInterface', newPrivateVirtualInterface_addressFamily - The address family for the BGP peer.

$sel:customerAddress:NewPrivateVirtualInterface', newPrivateVirtualInterface_customerAddress - The IP address assigned to the customer interface.

$sel:virtualInterfaceName:NewPrivateVirtualInterface', newPrivateVirtualInterface_virtualInterfaceName - The name of the virtual interface assigned by the customer network. The name has a maximum of 100 characters. The following are valid characters: a-z, 0-9 and a hyphen (-).

$sel:vlan:NewPrivateVirtualInterface', newPrivateVirtualInterface_vlan - The ID of the VLAN.

$sel:asn:NewPrivateVirtualInterface', newPrivateVirtualInterface_asn - The autonomous system (AS) number for Border Gateway Protocol (BGP) configuration.

The valid values are 1-2147483647.

newPrivateVirtualInterface_authKey :: Lens' NewPrivateVirtualInterface (Maybe Text)

The authentication key for BGP configuration. This string has a minimum length of 6 characters and and a maximun lenth of 80 characters.

newPrivateVirtualInterface_virtualGatewayId :: Lens' NewPrivateVirtualInterface (Maybe Text)

The ID of the virtual private gateway.

newPrivateVirtualInterface_mtu :: Lens' NewPrivateVirtualInterface (Maybe Int)

The maximum transmission unit (MTU), in bytes. The supported values are 1500 and 9001. The default value is 1500.

newPrivateVirtualInterface_tags :: Lens' NewPrivateVirtualInterface (Maybe (NonEmpty Tag))

The tags associated with the private virtual interface.

newPrivateVirtualInterface_directConnectGatewayId :: Lens' NewPrivateVirtualInterface (Maybe Text)

The ID of the Direct Connect gateway.

newPrivateVirtualInterface_amazonAddress :: Lens' NewPrivateVirtualInterface (Maybe Text)

The IP address assigned to the Amazon interface.

newPrivateVirtualInterface_addressFamily :: Lens' NewPrivateVirtualInterface (Maybe AddressFamily)

The address family for the BGP peer.

newPrivateVirtualInterface_customerAddress :: Lens' NewPrivateVirtualInterface (Maybe Text)

The IP address assigned to the customer interface.

newPrivateVirtualInterface_virtualInterfaceName :: Lens' NewPrivateVirtualInterface Text

The name of the virtual interface assigned by the customer network. The name has a maximum of 100 characters. The following are valid characters: a-z, 0-9 and a hyphen (-).

newPrivateVirtualInterface_vlan :: Lens' NewPrivateVirtualInterface Int

The ID of the VLAN.

newPrivateVirtualInterface_asn :: Lens' NewPrivateVirtualInterface Int

The autonomous system (AS) number for Border Gateway Protocol (BGP) configuration.

The valid values are 1-2147483647.