Network.AWS.DirectConnect.Types.NewPrivateVirtualInterfaceAllocation

Description

Documentation

data NewPrivateVirtualInterfaceAllocation

Information about a private virtual interface to be provisioned on a connection.

See: newNewPrivateVirtualInterfaceAllocation smart constructor.

Constructors

NewPrivateVirtualInterfaceAllocation'

Fields

authKey :: Maybe Text

The authentication key for BGP configuration. This string has a minimum length of 6 characters and and a maximun lenth of 80 characters.

mtu :: Maybe Int

The maximum transmission unit (MTU), in bytes. The supported values are 1500 and 9001. The default value is 1500.

tags :: Maybe (NonEmpty Tag)

The tags associated with the private virtual interface.

amazonAddress :: Maybe Text

The IP address assigned to the Amazon interface.

addressFamily :: Maybe AddressFamily

The address family for the BGP peer.

customerAddress :: Maybe Text

The IP address assigned to the customer interface.

virtualInterfaceName :: Text

The name of the virtual interface assigned by the customer network. The name has a maximum of 100 characters. The following are valid characters: a-z, 0-9 and a hyphen (-).

vlan :: Int

The ID of the VLAN.

asn :: Int

The autonomous system (AS) number for Border Gateway Protocol (BGP) configuration.

The valid values are 1-2147483647.

newNewPrivateVirtualInterfaceAllocation

Arguments

:: Text
$sel:virtualInterfaceName:NewPrivateVirtualInterfaceAllocation'

-> Int
$sel:vlan:NewPrivateVirtualInterfaceAllocation'

-> Int
$sel:asn:NewPrivateVirtualInterfaceAllocation'

-> NewPrivateVirtualInterfaceAllocation

Create a value of NewPrivateVirtualInterfaceAllocation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:authKey:NewPrivateVirtualInterfaceAllocation', newPrivateVirtualInterfaceAllocation_authKey - The authentication key for BGP configuration. This string has a minimum length of 6 characters and and a maximun lenth of 80 characters.

$sel:mtu:NewPrivateVirtualInterfaceAllocation', newPrivateVirtualInterfaceAllocation_mtu - The maximum transmission unit (MTU), in bytes. The supported values are 1500 and 9001. The default value is 1500.

$sel:tags:NewPrivateVirtualInterfaceAllocation', newPrivateVirtualInterfaceAllocation_tags - The tags associated with the private virtual interface.

$sel:amazonAddress:NewPrivateVirtualInterfaceAllocation', newPrivateVirtualInterfaceAllocation_amazonAddress - The IP address assigned to the Amazon interface.

$sel:addressFamily:NewPrivateVirtualInterfaceAllocation', newPrivateVirtualInterfaceAllocation_addressFamily - The address family for the BGP peer.

$sel:customerAddress:NewPrivateVirtualInterfaceAllocation', newPrivateVirtualInterfaceAllocation_customerAddress - The IP address assigned to the customer interface.

$sel:virtualInterfaceName:NewPrivateVirtualInterfaceAllocation', newPrivateVirtualInterfaceAllocation_virtualInterfaceName - The name of the virtual interface assigned by the customer network. The name has a maximum of 100 characters. The following are valid characters: a-z, 0-9 and a hyphen (-).

$sel:vlan:NewPrivateVirtualInterfaceAllocation', newPrivateVirtualInterfaceAllocation_vlan - The ID of the VLAN.

$sel:asn:NewPrivateVirtualInterfaceAllocation', newPrivateVirtualInterfaceAllocation_asn - The autonomous system (AS) number for Border Gateway Protocol (BGP) configuration.

The valid values are 1-2147483647.

newPrivateVirtualInterfaceAllocation_authKey :: Lens' NewPrivateVirtualInterfaceAllocation (Maybe Text)

The authentication key for BGP configuration. This string has a minimum length of 6 characters and and a maximun lenth of 80 characters.

newPrivateVirtualInterfaceAllocation_mtu :: Lens' NewPrivateVirtualInterfaceAllocation (Maybe Int)

The maximum transmission unit (MTU), in bytes. The supported values are 1500 and 9001. The default value is 1500.

newPrivateVirtualInterfaceAllocation_tags :: Lens' NewPrivateVirtualInterfaceAllocation (Maybe (NonEmpty Tag))

The tags associated with the private virtual interface.

newPrivateVirtualInterfaceAllocation_amazonAddress :: Lens' NewPrivateVirtualInterfaceAllocation (Maybe Text)

The IP address assigned to the Amazon interface.

newPrivateVirtualInterfaceAllocation_addressFamily :: Lens' NewPrivateVirtualInterfaceAllocation (Maybe AddressFamily)

The address family for the BGP peer.

newPrivateVirtualInterfaceAllocation_customerAddress :: Lens' NewPrivateVirtualInterfaceAllocation (Maybe Text)

The IP address assigned to the customer interface.

newPrivateVirtualInterfaceAllocation_virtualInterfaceName :: Lens' NewPrivateVirtualInterfaceAllocation Text

The name of the virtual interface assigned by the customer network. The name has a maximum of 100 characters. The following are valid characters: a-z, 0-9 and a hyphen (-).

newPrivateVirtualInterfaceAllocation_vlan :: Lens' NewPrivateVirtualInterfaceAllocation Int

The ID of the VLAN.

newPrivateVirtualInterfaceAllocation_asn :: Lens' NewPrivateVirtualInterfaceAllocation Int

The autonomous system (AS) number for Border Gateway Protocol (BGP) configuration.

The valid values are 1-2147483647.