Network.AWS.DirectConnect.Types.NewPublicVirtualInterface

Description

Documentation

data NewPublicVirtualInterface

Information about a public virtual interface.

See: newNewPublicVirtualInterface smart constructor.

Constructors

NewPublicVirtualInterface'

Fields

authKey :: Maybe Text

The authentication key for BGP configuration. This string has a minimum length of 6 characters and and a maximun lenth of 80 characters.

routeFilterPrefixes :: Maybe [RouteFilterPrefix]

The routes to be advertised to the Amazon Web Services network in this Region. Applies to public virtual interfaces.

tags :: Maybe (NonEmpty Tag)

The tags associated with the public virtual interface.

amazonAddress :: Maybe Text

The IP address assigned to the Amazon interface.

addressFamily :: Maybe AddressFamily

The address family for the BGP peer.

customerAddress :: Maybe Text

The IP address assigned to the customer interface.

virtualInterfaceName :: Text

The name of the virtual interface assigned by the customer network. The name has a maximum of 100 characters. The following are valid characters: a-z, 0-9 and a hyphen (-).

vlan :: Int

The ID of the VLAN.

asn :: Int

The autonomous system (AS) number for Border Gateway Protocol (BGP) configuration.

The valid values are 1-2147483647.

newNewPublicVirtualInterface

Arguments

:: Text
$sel:virtualInterfaceName:NewPublicVirtualInterface'

-> Int
$sel:vlan:NewPublicVirtualInterface'

-> Int
$sel:asn:NewPublicVirtualInterface'

-> NewPublicVirtualInterface

Create a value of NewPublicVirtualInterface with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:authKey:NewPublicVirtualInterface', newPublicVirtualInterface_authKey - The authentication key for BGP configuration. This string has a minimum length of 6 characters and and a maximun lenth of 80 characters.

$sel:routeFilterPrefixes:NewPublicVirtualInterface', newPublicVirtualInterface_routeFilterPrefixes - The routes to be advertised to the Amazon Web Services network in this Region. Applies to public virtual interfaces.

$sel:tags:NewPublicVirtualInterface', newPublicVirtualInterface_tags - The tags associated with the public virtual interface.

$sel:amazonAddress:NewPublicVirtualInterface', newPublicVirtualInterface_amazonAddress - The IP address assigned to the Amazon interface.

$sel:addressFamily:NewPublicVirtualInterface', newPublicVirtualInterface_addressFamily - The address family for the BGP peer.

$sel:customerAddress:NewPublicVirtualInterface', newPublicVirtualInterface_customerAddress - The IP address assigned to the customer interface.

$sel:virtualInterfaceName:NewPublicVirtualInterface', newPublicVirtualInterface_virtualInterfaceName - The name of the virtual interface assigned by the customer network. The name has a maximum of 100 characters. The following are valid characters: a-z, 0-9 and a hyphen (-).

$sel:vlan:NewPublicVirtualInterface', newPublicVirtualInterface_vlan - The ID of the VLAN.

$sel:asn:NewPublicVirtualInterface', newPublicVirtualInterface_asn - The autonomous system (AS) number for Border Gateway Protocol (BGP) configuration.

The valid values are 1-2147483647.

newPublicVirtualInterface_authKey :: Lens' NewPublicVirtualInterface (Maybe Text)

The authentication key for BGP configuration. This string has a minimum length of 6 characters and and a maximun lenth of 80 characters.

newPublicVirtualInterface_routeFilterPrefixes :: Lens' NewPublicVirtualInterface (Maybe [RouteFilterPrefix])

The routes to be advertised to the Amazon Web Services network in this Region. Applies to public virtual interfaces.

newPublicVirtualInterface_tags :: Lens' NewPublicVirtualInterface (Maybe (NonEmpty Tag))

The tags associated with the public virtual interface.

newPublicVirtualInterface_amazonAddress :: Lens' NewPublicVirtualInterface (Maybe Text)

The IP address assigned to the Amazon interface.

newPublicVirtualInterface_addressFamily :: Lens' NewPublicVirtualInterface (Maybe AddressFamily)

The address family for the BGP peer.

newPublicVirtualInterface_customerAddress :: Lens' NewPublicVirtualInterface (Maybe Text)

The IP address assigned to the customer interface.

newPublicVirtualInterface_virtualInterfaceName :: Lens' NewPublicVirtualInterface Text

The name of the virtual interface assigned by the customer network. The name has a maximum of 100 characters. The following are valid characters: a-z, 0-9 and a hyphen (-).

newPublicVirtualInterface_vlan :: Lens' NewPublicVirtualInterface Int

The ID of the VLAN.

newPublicVirtualInterface_asn :: Lens' NewPublicVirtualInterface Int

The autonomous system (AS) number for Border Gateway Protocol (BGP) configuration.

The valid values are 1-2147483647.