Network.AWS.DirectConnect.Types.NewPublicVirtualInterfaceAllocation

Description

Documentation

data NewPublicVirtualInterfaceAllocation

Information about a public virtual interface to be provisioned on a connection.

See: newNewPublicVirtualInterfaceAllocation smart constructor.

Constructors

NewPublicVirtualInterfaceAllocation'

Fields

authKey :: Maybe Text

The authentication key for BGP configuration. This string has a minimum length of 6 characters and and a maximun lenth of 80 characters.

routeFilterPrefixes :: Maybe [RouteFilterPrefix]

The routes to be advertised to the Amazon Web Services network in this Region. Applies to public virtual interfaces.

tags :: Maybe (NonEmpty Tag)

The tags associated with the public virtual interface.

amazonAddress :: Maybe Text

The IP address assigned to the Amazon interface.

addressFamily :: Maybe AddressFamily

The address family for the BGP peer.

customerAddress :: Maybe Text

The IP address assigned to the customer interface.

virtualInterfaceName :: Text

The name of the virtual interface assigned by the customer network. The name has a maximum of 100 characters. The following are valid characters: a-z, 0-9 and a hyphen (-).

vlan :: Int

The ID of the VLAN.

asn :: Int

The autonomous system (AS) number for Border Gateway Protocol (BGP) configuration.

The valid values are 1-2147483647.

newNewPublicVirtualInterfaceAllocation

Arguments

:: Text
$sel:virtualInterfaceName:NewPublicVirtualInterfaceAllocation'

-> Int
$sel:vlan:NewPublicVirtualInterfaceAllocation'

-> Int
$sel:asn:NewPublicVirtualInterfaceAllocation'

-> NewPublicVirtualInterfaceAllocation

Create a value of NewPublicVirtualInterfaceAllocation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:authKey:NewPublicVirtualInterfaceAllocation', newPublicVirtualInterfaceAllocation_authKey - The authentication key for BGP configuration. This string has a minimum length of 6 characters and and a maximun lenth of 80 characters.

$sel:routeFilterPrefixes:NewPublicVirtualInterfaceAllocation', newPublicVirtualInterfaceAllocation_routeFilterPrefixes - The routes to be advertised to the Amazon Web Services network in this Region. Applies to public virtual interfaces.

$sel:tags:NewPublicVirtualInterfaceAllocation', newPublicVirtualInterfaceAllocation_tags - The tags associated with the public virtual interface.

$sel:amazonAddress:NewPublicVirtualInterfaceAllocation', newPublicVirtualInterfaceAllocation_amazonAddress - The IP address assigned to the Amazon interface.

$sel:addressFamily:NewPublicVirtualInterfaceAllocation', newPublicVirtualInterfaceAllocation_addressFamily - The address family for the BGP peer.

$sel:customerAddress:NewPublicVirtualInterfaceAllocation', newPublicVirtualInterfaceAllocation_customerAddress - The IP address assigned to the customer interface.

$sel:virtualInterfaceName:NewPublicVirtualInterfaceAllocation', newPublicVirtualInterfaceAllocation_virtualInterfaceName - The name of the virtual interface assigned by the customer network. The name has a maximum of 100 characters. The following are valid characters: a-z, 0-9 and a hyphen (-).

$sel:vlan:NewPublicVirtualInterfaceAllocation', newPublicVirtualInterfaceAllocation_vlan - The ID of the VLAN.

$sel:asn:NewPublicVirtualInterfaceAllocation', newPublicVirtualInterfaceAllocation_asn - The autonomous system (AS) number for Border Gateway Protocol (BGP) configuration.

The valid values are 1-2147483647.

newPublicVirtualInterfaceAllocation_authKey :: Lens' NewPublicVirtualInterfaceAllocation (Maybe Text)

The authentication key for BGP configuration. This string has a minimum length of 6 characters and and a maximun lenth of 80 characters.

newPublicVirtualInterfaceAllocation_routeFilterPrefixes :: Lens' NewPublicVirtualInterfaceAllocation (Maybe [RouteFilterPrefix])

The routes to be advertised to the Amazon Web Services network in this Region. Applies to public virtual interfaces.

newPublicVirtualInterfaceAllocation_tags :: Lens' NewPublicVirtualInterfaceAllocation (Maybe (NonEmpty Tag))

The tags associated with the public virtual interface.

newPublicVirtualInterfaceAllocation_amazonAddress :: Lens' NewPublicVirtualInterfaceAllocation (Maybe Text)

The IP address assigned to the Amazon interface.

newPublicVirtualInterfaceAllocation_addressFamily :: Lens' NewPublicVirtualInterfaceAllocation (Maybe AddressFamily)

The address family for the BGP peer.

newPublicVirtualInterfaceAllocation_customerAddress :: Lens' NewPublicVirtualInterfaceAllocation (Maybe Text)

The IP address assigned to the customer interface.

newPublicVirtualInterfaceAllocation_virtualInterfaceName :: Lens' NewPublicVirtualInterfaceAllocation Text

The name of the virtual interface assigned by the customer network. The name has a maximum of 100 characters. The following are valid characters: a-z, 0-9 and a hyphen (-).

newPublicVirtualInterfaceAllocation_vlan :: Lens' NewPublicVirtualInterfaceAllocation Int

The ID of the VLAN.

newPublicVirtualInterfaceAllocation_asn :: Lens' NewPublicVirtualInterfaceAllocation Int

The autonomous system (AS) number for Border Gateway Protocol (BGP) configuration.

The valid values are 1-2147483647.