Network.AWS.DirectConnect.Types.NewTransitVirtualInterface

Description

Documentation

data NewTransitVirtualInterface

Information about a transit virtual interface.

See: newNewTransitVirtualInterface smart constructor.

Constructors

NewTransitVirtualInterface'

Fields

authKey :: Maybe Text

The authentication key for BGP configuration. This string has a minimum length of 6 characters and and a maximun lenth of 80 characters.

asn :: Maybe Int

The autonomous system (AS) number for Border Gateway Protocol (BGP) configuration.

The valid values are 1-2147483647.

mtu :: Maybe Int

The maximum transmission unit (MTU), in bytes. The supported values are 1500 and 9001. The default value is 1500.

tags :: Maybe (NonEmpty Tag)

The tags associated with the transitive virtual interface.

directConnectGatewayId :: Maybe Text

The ID of the Direct Connect gateway.

virtualInterfaceName :: Maybe Text

The name of the virtual interface assigned by the customer network. The name has a maximum of 100 characters. The following are valid characters: a-z, 0-9 and a hyphen (-).

amazonAddress :: Maybe Text

The IP address assigned to the Amazon interface.

addressFamily :: Maybe AddressFamily

The address family for the BGP peer.

vlan :: Maybe Int

The ID of the VLAN.

customerAddress :: Maybe Text

The IP address assigned to the customer interface.

newNewTransitVirtualInterface :: NewTransitVirtualInterface

Create a value of NewTransitVirtualInterface with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:authKey:NewTransitVirtualInterface', newTransitVirtualInterface_authKey - The authentication key for BGP configuration. This string has a minimum length of 6 characters and and a maximun lenth of 80 characters.

$sel:asn:NewTransitVirtualInterface', newTransitVirtualInterface_asn - The autonomous system (AS) number for Border Gateway Protocol (BGP) configuration.

The valid values are 1-2147483647.

$sel:mtu:NewTransitVirtualInterface', newTransitVirtualInterface_mtu - The maximum transmission unit (MTU), in bytes. The supported values are 1500 and 9001. The default value is 1500.

$sel:tags:NewTransitVirtualInterface', newTransitVirtualInterface_tags - The tags associated with the transitive virtual interface.

$sel:directConnectGatewayId:NewTransitVirtualInterface', newTransitVirtualInterface_directConnectGatewayId - The ID of the Direct Connect gateway.

$sel:virtualInterfaceName:NewTransitVirtualInterface', newTransitVirtualInterface_virtualInterfaceName - The name of the virtual interface assigned by the customer network. The name has a maximum of 100 characters. The following are valid characters: a-z, 0-9 and a hyphen (-).

$sel:amazonAddress:NewTransitVirtualInterface', newTransitVirtualInterface_amazonAddress - The IP address assigned to the Amazon interface.

$sel:addressFamily:NewTransitVirtualInterface', newTransitVirtualInterface_addressFamily - The address family for the BGP peer.

$sel:vlan:NewTransitVirtualInterface', newTransitVirtualInterface_vlan - The ID of the VLAN.

$sel:customerAddress:NewTransitVirtualInterface', newTransitVirtualInterface_customerAddress - The IP address assigned to the customer interface.

newTransitVirtualInterface_authKey :: Lens' NewTransitVirtualInterface (Maybe Text)

The authentication key for BGP configuration. This string has a minimum length of 6 characters and and a maximun lenth of 80 characters.

newTransitVirtualInterface_asn :: Lens' NewTransitVirtualInterface (Maybe Int)

The autonomous system (AS) number for Border Gateway Protocol (BGP) configuration.

The valid values are 1-2147483647.

newTransitVirtualInterface_mtu :: Lens' NewTransitVirtualInterface (Maybe Int)

The maximum transmission unit (MTU), in bytes. The supported values are 1500 and 9001. The default value is 1500.

newTransitVirtualInterface_tags :: Lens' NewTransitVirtualInterface (Maybe (NonEmpty Tag))

The tags associated with the transitive virtual interface.

newTransitVirtualInterface_directConnectGatewayId :: Lens' NewTransitVirtualInterface (Maybe Text)

The ID of the Direct Connect gateway.

newTransitVirtualInterface_virtualInterfaceName :: Lens' NewTransitVirtualInterface (Maybe Text)

The name of the virtual interface assigned by the customer network. The name has a maximum of 100 characters. The following are valid characters: a-z, 0-9 and a hyphen (-).

newTransitVirtualInterface_amazonAddress :: Lens' NewTransitVirtualInterface (Maybe Text)

The IP address assigned to the Amazon interface.

newTransitVirtualInterface_addressFamily :: Lens' NewTransitVirtualInterface (Maybe AddressFamily)

The address family for the BGP peer.

newTransitVirtualInterface_vlan :: Lens' NewTransitVirtualInterface (Maybe Int)

The ID of the VLAN.

newTransitVirtualInterface_customerAddress :: Lens' NewTransitVirtualInterface (Maybe Text)

The IP address assigned to the customer interface.