Network.AWS.DirectConnect.UpdateDirectConnectGatewayAssociation

Description

Updates the specified attributes of the Direct Connect gateway association.

Add or remove prefixes from the association.

Creating a Request

data UpdateDirectConnectGatewayAssociation

See: newUpdateDirectConnectGatewayAssociation smart constructor.

Constructors

UpdateDirectConnectGatewayAssociation'

Fields

removeAllowedPrefixesToDirectConnectGateway :: Maybe [RouteFilterPrefix]

The Amazon VPC prefixes to no longer advertise to the Direct Connect gateway.

addAllowedPrefixesToDirectConnectGateway :: Maybe [RouteFilterPrefix]

The Amazon VPC prefixes to advertise to the Direct Connect gateway.

associationId :: Maybe Text

The ID of the Direct Connect gateway association.

newUpdateDirectConnectGatewayAssociation :: UpdateDirectConnectGatewayAssociation

Create a value of UpdateDirectConnectGatewayAssociation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:removeAllowedPrefixesToDirectConnectGateway:UpdateDirectConnectGatewayAssociation', updateDirectConnectGatewayAssociation_removeAllowedPrefixesToDirectConnectGateway - The Amazon VPC prefixes to no longer advertise to the Direct Connect gateway.

$sel:addAllowedPrefixesToDirectConnectGateway:UpdateDirectConnectGatewayAssociation', updateDirectConnectGatewayAssociation_addAllowedPrefixesToDirectConnectGateway - The Amazon VPC prefixes to advertise to the Direct Connect gateway.

$sel:associationId:UpdateDirectConnectGatewayAssociation', updateDirectConnectGatewayAssociation_associationId - The ID of the Direct Connect gateway association.

Request Lenses

updateDirectConnectGatewayAssociation_removeAllowedPrefixesToDirectConnectGateway :: Lens' UpdateDirectConnectGatewayAssociation (Maybe [RouteFilterPrefix])

The Amazon VPC prefixes to no longer advertise to the Direct Connect gateway.

updateDirectConnectGatewayAssociation_addAllowedPrefixesToDirectConnectGateway :: Lens' UpdateDirectConnectGatewayAssociation (Maybe [RouteFilterPrefix])

The Amazon VPC prefixes to advertise to the Direct Connect gateway.

updateDirectConnectGatewayAssociation_associationId :: Lens' UpdateDirectConnectGatewayAssociation (Maybe Text)

The ID of the Direct Connect gateway association.

Destructuring the Response

data UpdateDirectConnectGatewayAssociationResponse

See: newUpdateDirectConnectGatewayAssociationResponse smart constructor.

Constructors

UpdateDirectConnectGatewayAssociationResponse'

Fields

directConnectGatewayAssociation :: Maybe DirectConnectGatewayAssociation

httpStatus :: Int

The response's http status code.

newUpdateDirectConnectGatewayAssociationResponse

Arguments

:: Int
$sel:httpStatus:UpdateDirectConnectGatewayAssociationResponse'

-> UpdateDirectConnectGatewayAssociationResponse

Create a value of UpdateDirectConnectGatewayAssociationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:directConnectGatewayAssociation:UpdateDirectConnectGatewayAssociationResponse', updateDirectConnectGatewayAssociationResponse_directConnectGatewayAssociation - Undocumented member.

$sel:httpStatus:UpdateDirectConnectGatewayAssociationResponse', updateDirectConnectGatewayAssociationResponse_httpStatus - The response's http status code.

Response Lenses

updateDirectConnectGatewayAssociationResponse_directConnectGatewayAssociation :: Lens' UpdateDirectConnectGatewayAssociationResponse (Maybe DirectConnectGatewayAssociation)

Undocumented member.

updateDirectConnectGatewayAssociationResponse_httpStatus :: Lens' UpdateDirectConnectGatewayAssociationResponse Int

The response's http status code.