Network.AWS.DirectConnect.V2012_10_25.DescribeVirtualInterfaces

Description

Displays all virtual interfaces for an AWS account. Virtual interfaces deleted fewer than 15 minutes before DescribeVirtualInterfaces is called are also returned. If a connection ID is included then only virtual interfaces associated with this connection will be returned. If a virtual interface ID is included then only a single virtual interface will be returned. A virtual interface (VLAN) transmits the traffic between the AWS Direct Connect location and the customer. If a connection ID is provided, only virtual interfaces provisioned on the specified connection will be returned. If a virtual interface ID is provided, only this particular virtual interface will be returned.

Documentation

describeVirtualInterfaces :: DescribeVirtualInterfaces

Minimum specification for a DescribeVirtualInterfaces request.

data DescribeVirtualInterfaces

Constructors

DescribeVirtualInterfaces

Fields

_dvirConnectionId :: Maybe Text

ID of the connection. Example: dxcon-fg5678gh Default: None.

_dvirVirtualInterfaceId :: Maybe Text

ID of the virtual interface. Example: dxvif-123dfg56 Default: None.

dvirVirtualInterfaceId :: Lens' DescribeVirtualInterfaces (Maybe Text)

dvirConnectionId :: Lens' DescribeVirtualInterfaces (Maybe Text)

data DescribeVirtualInterfacesResponse

Constructors

DescribeVirtualInterfacesResponse

Fields

_vkVirtualInterfaces :: [VirtualInterface]

A list of virtual interfaces.

vkVirtualInterfaces :: Iso' DescribeVirtualInterfacesResponse [VirtualInterface]