Network.AWS.DirectoryService.AddIpRoutes

Description

If the DNS server for your self-managed domain uses a publicly addressable IP address, you must add a CIDR address block to correctly route traffic to and from your Microsoft AD on Amazon Web Services. AddIpRoutes adds this address block. You can also use AddIpRoutes to facilitate routing traffic that uses public IP ranges from your Microsoft AD on Amazon Web Services to a peer VPC.

Before you call AddIpRoutes, ensure that all of the required permissions have been explicitly granted through a policy. For details about what permissions are required to run the AddIpRoutes operation, see Directory Service API Permissions: Actions, Resources, and Conditions Reference.

Creating a Request

data AddIpRoutes

See: newAddIpRoutes smart constructor.

Constructors

AddIpRoutes'

Fields

updateSecurityGroupForDirectoryControllers :: Maybe Bool

If set to true, updates the inbound and outbound rules of the security group that has the description: "Amazon Web Services created security group for directory ID directory controllers." Following are the new rules:

Inbound:

Type: Custom UDP Rule, Protocol: UDP, Range: 88, Source: 0.0.0.0/0

Type: Custom UDP Rule, Protocol: UDP, Range: 123, Source: 0.0.0.0/0

Type: Custom UDP Rule, Protocol: UDP, Range: 138, Source: 0.0.0.0/0

Type: Custom UDP Rule, Protocol: UDP, Range: 389, Source: 0.0.0.0/0

Type: Custom UDP Rule, Protocol: UDP, Range: 464, Source: 0.0.0.0/0

Type: Custom UDP Rule, Protocol: UDP, Range: 445, Source: 0.0.0.0/0

Type: Custom TCP Rule, Protocol: TCP, Range: 88, Source: 0.0.0.0/0

Type: Custom TCP Rule, Protocol: TCP, Range: 135, Source: 0.0.0.0/0

Type: Custom TCP Rule, Protocol: TCP, Range: 445, Source: 0.0.0.0/0

Type: Custom TCP Rule, Protocol: TCP, Range: 464, Source: 0.0.0.0/0

Type: Custom TCP Rule, Protocol: TCP, Range: 636, Source: 0.0.0.0/0

Type: Custom TCP Rule, Protocol: TCP, Range: 1024-65535, Source: 0.0.0.0/0

Type: Custom TCP Rule, Protocol: TCP, Range: 3268-33269, Source: 0.0.0.0/0

Type: DNS (UDP), Protocol: UDP, Range: 53, Source: 0.0.0.0/0

Type: DNS (TCP), Protocol: TCP, Range: 53, Source: 0.0.0.0/0

Type: LDAP, Protocol: TCP, Range: 389, Source: 0.0.0.0/0

Type: All ICMP, Protocol: All, Range: N/A, Source: 0.0.0.0/0

Outbound:

Type: All traffic, Protocol: All, Range: All, Destination: 0.0.0.0/0

These security rules impact an internal network interface that is not exposed publicly.

directoryId :: Text

Identifier (ID) of the directory to which to add the address block.

ipRoutes :: [IpRoute]

IP address blocks, using CIDR format, of the traffic to route. This is often the IP address block of the DNS server used for your self-managed domain.

newAddIpRoutes

Arguments

:: Text
$sel:directoryId:AddIpRoutes'

-> AddIpRoutes

Create a value of AddIpRoutes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:updateSecurityGroupForDirectoryControllers:AddIpRoutes', addIpRoutes_updateSecurityGroupForDirectoryControllers - If set to true, updates the inbound and outbound rules of the security group that has the description: "Amazon Web Services created security group for directory ID directory controllers." Following are the new rules:

Inbound:

Type: Custom UDP Rule, Protocol: UDP, Range: 88, Source: 0.0.0.0/0

Type: Custom UDP Rule, Protocol: UDP, Range: 123, Source: 0.0.0.0/0

Type: Custom UDP Rule, Protocol: UDP, Range: 138, Source: 0.0.0.0/0

Type: Custom UDP Rule, Protocol: UDP, Range: 389, Source: 0.0.0.0/0

Type: Custom UDP Rule, Protocol: UDP, Range: 464, Source: 0.0.0.0/0

Type: Custom UDP Rule, Protocol: UDP, Range: 445, Source: 0.0.0.0/0

Type: Custom TCP Rule, Protocol: TCP, Range: 88, Source: 0.0.0.0/0

Type: Custom TCP Rule, Protocol: TCP, Range: 135, Source: 0.0.0.0/0

Type: Custom TCP Rule, Protocol: TCP, Range: 445, Source: 0.0.0.0/0

Type: Custom TCP Rule, Protocol: TCP, Range: 464, Source: 0.0.0.0/0

Type: Custom TCP Rule, Protocol: TCP, Range: 636, Source: 0.0.0.0/0

Type: Custom TCP Rule, Protocol: TCP, Range: 1024-65535, Source: 0.0.0.0/0

Type: Custom TCP Rule, Protocol: TCP, Range: 3268-33269, Source: 0.0.0.0/0

Type: DNS (UDP), Protocol: UDP, Range: 53, Source: 0.0.0.0/0

Type: DNS (TCP), Protocol: TCP, Range: 53, Source: 0.0.0.0/0

Type: LDAP, Protocol: TCP, Range: 389, Source: 0.0.0.0/0

Type: All ICMP, Protocol: All, Range: N/A, Source: 0.0.0.0/0

Outbound:

Type: All traffic, Protocol: All, Range: All, Destination: 0.0.0.0/0

These security rules impact an internal network interface that is not exposed publicly.

$sel:directoryId:AddIpRoutes', addIpRoutes_directoryId - Identifier (ID) of the directory to which to add the address block.

$sel:ipRoutes:AddIpRoutes', addIpRoutes_ipRoutes - IP address blocks, using CIDR format, of the traffic to route. This is often the IP address block of the DNS server used for your self-managed domain.

Request Lenses

addIpRoutes_updateSecurityGroupForDirectoryControllers :: Lens' AddIpRoutes (Maybe Bool)

If set to true, updates the inbound and outbound rules of the security group that has the description: "Amazon Web Services created security group for directory ID directory controllers." Following are the new rules:

Inbound:

Type: Custom UDP Rule, Protocol: UDP, Range: 88, Source: 0.0.0.0/0

Type: Custom UDP Rule, Protocol: UDP, Range: 123, Source: 0.0.0.0/0

Type: Custom UDP Rule, Protocol: UDP, Range: 138, Source: 0.0.0.0/0

Type: Custom UDP Rule, Protocol: UDP, Range: 389, Source: 0.0.0.0/0

Type: Custom UDP Rule, Protocol: UDP, Range: 464, Source: 0.0.0.0/0

Type: Custom UDP Rule, Protocol: UDP, Range: 445, Source: 0.0.0.0/0

Type: Custom TCP Rule, Protocol: TCP, Range: 88, Source: 0.0.0.0/0

Type: Custom TCP Rule, Protocol: TCP, Range: 135, Source: 0.0.0.0/0

Type: Custom TCP Rule, Protocol: TCP, Range: 445, Source: 0.0.0.0/0

Type: Custom TCP Rule, Protocol: TCP, Range: 464, Source: 0.0.0.0/0

Type: Custom TCP Rule, Protocol: TCP, Range: 636, Source: 0.0.0.0/0

Type: Custom TCP Rule, Protocol: TCP, Range: 1024-65535, Source: 0.0.0.0/0

Type: Custom TCP Rule, Protocol: TCP, Range: 3268-33269, Source: 0.0.0.0/0

Type: DNS (UDP), Protocol: UDP, Range: 53, Source: 0.0.0.0/0

Type: DNS (TCP), Protocol: TCP, Range: 53, Source: 0.0.0.0/0

Type: LDAP, Protocol: TCP, Range: 389, Source: 0.0.0.0/0

Type: All ICMP, Protocol: All, Range: N/A, Source: 0.0.0.0/0

Outbound:

Type: All traffic, Protocol: All, Range: All, Destination: 0.0.0.0/0

These security rules impact an internal network interface that is not exposed publicly.

addIpRoutes_directoryId :: Lens' AddIpRoutes Text

Identifier (ID) of the directory to which to add the address block.

addIpRoutes_ipRoutes :: Lens' AddIpRoutes [IpRoute]

IP address blocks, using CIDR format, of the traffic to route. This is often the IP address block of the DNS server used for your self-managed domain.

Destructuring the Response

data AddIpRoutesResponse

See: newAddIpRoutesResponse smart constructor.

Constructors

AddIpRoutesResponse'

Fields

httpStatus :: Int

The response's http status code.

newAddIpRoutesResponse

Arguments

:: Int
$sel:httpStatus:AddIpRoutesResponse'

-> AddIpRoutesResponse

Create a value of AddIpRoutesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:AddIpRoutesResponse', addIpRoutesResponse_httpStatus - The response's http status code.

Response Lenses

addIpRoutesResponse_httpStatus :: Lens' AddIpRoutesResponse Int

The response's http status code.