Network.AWS.DirectoryService.ConnectDirectory

Description

Creates an AD Connector to connect to a self-managed directory.

Before you call ConnectDirectory, ensure that all of the required permissions have been explicitly granted through a policy. For details about what permissions are required to run the ConnectDirectory operation, see Directory Service API Permissions: Actions, Resources, and Conditions Reference.

Creating a Request

data ConnectDirectory

Contains the inputs for the ConnectDirectory operation.

See: newConnectDirectory smart constructor.

Constructors

ConnectDirectory'

Fields

shortName :: Maybe Text

The NetBIOS name of your self-managed directory, such as CORP.

description :: Maybe Text

A description for the directory.

tags :: Maybe [Tag]

The tags to be assigned to AD Connector.

name :: Text

The fully qualified name of your self-managed directory, such as corp.example.com.

password :: Sensitive Text

The password for your self-managed user account.

size :: DirectorySize

The size of the directory.

connectSettings :: DirectoryConnectSettings

A DirectoryConnectSettings object that contains additional information for the operation.

newConnectDirectory

Arguments

:: Text
$sel:name:ConnectDirectory'

-> Text
$sel:password:ConnectDirectory'

-> DirectorySize
$sel:size:ConnectDirectory'

-> DirectoryConnectSettings
$sel:connectSettings:ConnectDirectory'

-> ConnectDirectory

Create a value of ConnectDirectory with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:shortName:ConnectDirectory', connectDirectory_shortName - The NetBIOS name of your self-managed directory, such as CORP.

$sel:description:ConnectDirectory', connectDirectory_description - A description for the directory.

$sel:tags:ConnectDirectory', connectDirectory_tags - The tags to be assigned to AD Connector.

$sel:name:ConnectDirectory', connectDirectory_name - The fully qualified name of your self-managed directory, such as corp.example.com.

$sel:password:ConnectDirectory', connectDirectory_password - The password for your self-managed user account.

$sel:size:ConnectDirectory', connectDirectory_size - The size of the directory.

$sel:connectSettings:ConnectDirectory', connectDirectory_connectSettings - A DirectoryConnectSettings object that contains additional information for the operation.

Request Lenses

connectDirectory_shortName :: Lens' ConnectDirectory (Maybe Text)

The NetBIOS name of your self-managed directory, such as CORP.

connectDirectory_description :: Lens' ConnectDirectory (Maybe Text)

A description for the directory.

connectDirectory_tags :: Lens' ConnectDirectory (Maybe [Tag])

The tags to be assigned to AD Connector.

connectDirectory_name :: Lens' ConnectDirectory Text

The fully qualified name of your self-managed directory, such as corp.example.com.

connectDirectory_password :: Lens' ConnectDirectory Text

The password for your self-managed user account.

connectDirectory_size :: Lens' ConnectDirectory DirectorySize

The size of the directory.

connectDirectory_connectSettings :: Lens' ConnectDirectory DirectoryConnectSettings

A DirectoryConnectSettings object that contains additional information for the operation.

Destructuring the Response

data ConnectDirectoryResponse

Contains the results of the ConnectDirectory operation.

See: newConnectDirectoryResponse smart constructor.

Constructors

ConnectDirectoryResponse'

Fields

directoryId :: Maybe Text

The identifier of the new directory.

httpStatus :: Int

The response's http status code.

newConnectDirectoryResponse

Arguments

:: Int
$sel:httpStatus:ConnectDirectoryResponse'

-> ConnectDirectoryResponse

Create a value of ConnectDirectoryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:directoryId:ConnectDirectoryResponse', connectDirectoryResponse_directoryId - The identifier of the new directory.

$sel:httpStatus:ConnectDirectoryResponse', connectDirectoryResponse_httpStatus - The response's http status code.

Response Lenses

connectDirectoryResponse_directoryId :: Lens' ConnectDirectoryResponse (Maybe Text)

The identifier of the new directory.

connectDirectoryResponse_httpStatus :: Lens' ConnectDirectoryResponse Int

The response's http status code.