Network.AWS.DirectoryService.CreateTrust

Description

Directory Service for Microsoft Active Directory allows you to configure trust relationships. For example, you can establish a trust between your Managed Microsoft AD directory, and your existing self-managed Microsoft Active Directory. This would allow you to provide users and groups access to resources in either domain, with a single set of credentials.

This action initiates the creation of the Amazon Web Services side of a trust relationship between an Managed Microsoft AD directory and an external domain. You can create either a forest trust or an external trust.

Creating a Request

data CreateTrust

Directory Service for Microsoft Active Directory allows you to configure trust relationships. For example, you can establish a trust between your Managed Microsoft AD directory, and your existing self-managed Microsoft Active Directory. This would allow you to provide users and groups access to resources in either domain, with a single set of credentials.

This action initiates the creation of the Amazon Web Services side of a trust relationship between an Managed Microsoft AD directory and an external domain.

See: newCreateTrust smart constructor.

Constructors

CreateTrust'

Fields

trustType :: Maybe TrustType

The trust relationship type. Forest is the default.

selectiveAuth :: Maybe SelectiveAuth

Optional parameter to enable selective authentication for the trust.

conditionalForwarderIpAddrs :: Maybe [Text]

The IP addresses of the remote DNS server associated with RemoteDomainName.

directoryId :: Text

The Directory ID of the Managed Microsoft AD directory for which to establish the trust relationship.

remoteDomainName :: Text

The Fully Qualified Domain Name (FQDN) of the external domain for which to create the trust relationship.

trustPassword :: Sensitive Text

The trust password. The must be the same password that was used when creating the trust relationship on the external domain.

trustDirection :: TrustDirection

The direction of the trust relationship.

newCreateTrust

Arguments

:: Text
$sel:directoryId:CreateTrust'

-> Text
$sel:remoteDomainName:CreateTrust'

-> Text
$sel:trustPassword:CreateTrust'

-> TrustDirection
$sel:trustDirection:CreateTrust'

-> CreateTrust

Create a value of CreateTrust with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:trustType:CreateTrust', createTrust_trustType - The trust relationship type. Forest is the default.

$sel:selectiveAuth:CreateTrust', createTrust_selectiveAuth - Optional parameter to enable selective authentication for the trust.

$sel:conditionalForwarderIpAddrs:CreateTrust', createTrust_conditionalForwarderIpAddrs - The IP addresses of the remote DNS server associated with RemoteDomainName.

$sel:directoryId:CreateTrust', createTrust_directoryId - The Directory ID of the Managed Microsoft AD directory for which to establish the trust relationship.

$sel:remoteDomainName:CreateTrust', createTrust_remoteDomainName - The Fully Qualified Domain Name (FQDN) of the external domain for which to create the trust relationship.

$sel:trustPassword:CreateTrust', createTrust_trustPassword - The trust password. The must be the same password that was used when creating the trust relationship on the external domain.

$sel:trustDirection:CreateTrust', createTrust_trustDirection - The direction of the trust relationship.

Request Lenses

createTrust_trustType :: Lens' CreateTrust (Maybe TrustType)

The trust relationship type. Forest is the default.

createTrust_selectiveAuth :: Lens' CreateTrust (Maybe SelectiveAuth)

Optional parameter to enable selective authentication for the trust.

createTrust_conditionalForwarderIpAddrs :: Lens' CreateTrust (Maybe [Text])

The IP addresses of the remote DNS server associated with RemoteDomainName.

createTrust_directoryId :: Lens' CreateTrust Text

The Directory ID of the Managed Microsoft AD directory for which to establish the trust relationship.

createTrust_remoteDomainName :: Lens' CreateTrust Text

The Fully Qualified Domain Name (FQDN) of the external domain for which to create the trust relationship.

createTrust_trustPassword :: Lens' CreateTrust Text

The trust password. The must be the same password that was used when creating the trust relationship on the external domain.

createTrust_trustDirection :: Lens' CreateTrust TrustDirection

The direction of the trust relationship.

Destructuring the Response

data CreateTrustResponse

The result of a CreateTrust request.

See: newCreateTrustResponse smart constructor.

Constructors

CreateTrustResponse'

Fields

trustId :: Maybe Text

A unique identifier for the trust relationship that was created.

httpStatus :: Int

The response's http status code.

newCreateTrustResponse

Arguments

:: Int
$sel:httpStatus:CreateTrustResponse'

-> CreateTrustResponse

Create a value of CreateTrustResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:trustId:CreateTrustResponse', createTrustResponse_trustId - A unique identifier for the trust relationship that was created.

$sel:httpStatus:CreateTrustResponse', createTrustResponse_httpStatus - The response's http status code.

Response Lenses

createTrustResponse_trustId :: Lens' CreateTrustResponse (Maybe Text)

A unique identifier for the trust relationship that was created.

createTrustResponse_httpStatus :: Lens' CreateTrustResponse Int

The response's http status code.