Network.AWS.DirectoryService.DescribeClientAuthenticationSettings

Description

Retrieves information about the type of client authentication for the specified directory, if the type is specified. If no type is specified, information about all client authentication types that are supported for the specified directory is retrieved. Currently, only SmartCard is supported.

Creating a Request

data DescribeClientAuthenticationSettings

See: newDescribeClientAuthenticationSettings smart constructor.

Constructors

DescribeClientAuthenticationSettings'

Fields

nextToken :: Maybe Text

The DescribeClientAuthenticationSettingsResult.NextToken value from a previous call to DescribeClientAuthenticationSettings. Pass null if this is the first call.

limit :: Maybe Natural

The maximum number of items to return. If this value is zero, the maximum number of items is specified by the limitations of the operation.

type' :: Maybe ClientAuthenticationType

The type of client authentication for which to retrieve information. If no type is specified, a list of all client authentication types that are supported for the specified directory is retrieved.

directoryId :: Text

The identifier of the directory for which to retrieve information.

newDescribeClientAuthenticationSettings

Arguments

:: Text
$sel:directoryId:DescribeClientAuthenticationSettings'

-> DescribeClientAuthenticationSettings

Create a value of DescribeClientAuthenticationSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeClientAuthenticationSettings', describeClientAuthenticationSettings_nextToken - The DescribeClientAuthenticationSettingsResult.NextToken value from a previous call to DescribeClientAuthenticationSettings. Pass null if this is the first call.

$sel:limit:DescribeClientAuthenticationSettings', describeClientAuthenticationSettings_limit - The maximum number of items to return. If this value is zero, the maximum number of items is specified by the limitations of the operation.

$sel:type':DescribeClientAuthenticationSettings', describeClientAuthenticationSettings_type - The type of client authentication for which to retrieve information. If no type is specified, a list of all client authentication types that are supported for the specified directory is retrieved.

$sel:directoryId:DescribeClientAuthenticationSettings', describeClientAuthenticationSettings_directoryId - The identifier of the directory for which to retrieve information.

Request Lenses

describeClientAuthenticationSettings_nextToken :: Lens' DescribeClientAuthenticationSettings (Maybe Text)

The DescribeClientAuthenticationSettingsResult.NextToken value from a previous call to DescribeClientAuthenticationSettings. Pass null if this is the first call.

describeClientAuthenticationSettings_limit :: Lens' DescribeClientAuthenticationSettings (Maybe Natural)

The maximum number of items to return. If this value is zero, the maximum number of items is specified by the limitations of the operation.

describeClientAuthenticationSettings_type :: Lens' DescribeClientAuthenticationSettings (Maybe ClientAuthenticationType)

The type of client authentication for which to retrieve information. If no type is specified, a list of all client authentication types that are supported for the specified directory is retrieved.

describeClientAuthenticationSettings_directoryId :: Lens' DescribeClientAuthenticationSettings Text

The identifier of the directory for which to retrieve information.

Destructuring the Response

data DescribeClientAuthenticationSettingsResponse

See: newDescribeClientAuthenticationSettingsResponse smart constructor.

Constructors

DescribeClientAuthenticationSettingsResponse'

Fields

nextToken :: Maybe Text

The next token used to retrieve the client authentication settings if the number of setting types exceeds page limit and there is another page.

clientAuthenticationSettingsInfo :: Maybe [ClientAuthenticationSettingInfo]

Information about the type of client authentication for the specified directory. The following information is retrieved: The date and time when the status of the client authentication type was last updated, whether the client authentication type is enabled or disabled, and the type of client authentication.

httpStatus :: Int

The response's http status code.

newDescribeClientAuthenticationSettingsResponse

Arguments

:: Int
$sel:httpStatus:DescribeClientAuthenticationSettingsResponse'

-> DescribeClientAuthenticationSettingsResponse

Create a value of DescribeClientAuthenticationSettingsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeClientAuthenticationSettings', describeClientAuthenticationSettingsResponse_nextToken - The next token used to retrieve the client authentication settings if the number of setting types exceeds page limit and there is another page.

$sel:clientAuthenticationSettingsInfo:DescribeClientAuthenticationSettingsResponse', describeClientAuthenticationSettingsResponse_clientAuthenticationSettingsInfo - Information about the type of client authentication for the specified directory. The following information is retrieved: The date and time when the status of the client authentication type was last updated, whether the client authentication type is enabled or disabled, and the type of client authentication.

$sel:httpStatus:DescribeClientAuthenticationSettingsResponse', describeClientAuthenticationSettingsResponse_httpStatus - The response's http status code.

Response Lenses

describeClientAuthenticationSettingsResponse_nextToken :: Lens' DescribeClientAuthenticationSettingsResponse (Maybe Text)

The next token used to retrieve the client authentication settings if the number of setting types exceeds page limit and there is another page.

describeClientAuthenticationSettingsResponse_clientAuthenticationSettingsInfo :: Lens' DescribeClientAuthenticationSettingsResponse (Maybe [ClientAuthenticationSettingInfo])

Information about the type of client authentication for the specified directory. The following information is retrieved: The date and time when the status of the client authentication type was last updated, whether the client authentication type is enabled or disabled, and the type of client authentication.

describeClientAuthenticationSettingsResponse_httpStatus :: Lens' DescribeClientAuthenticationSettingsResponse Int

The response's http status code.