Network.AWS.DirectoryService.DescribeDirectories

Description

Obtains information about the directories that belong to this account.

You can retrieve information about specific directories by passing the directory identifiers in the DirectoryIds parameter. Otherwise, all directories that belong to the current account are returned.

This operation supports pagination with the use of the NextToken request and response parameters. If more results are available, the DescribeDirectoriesResult.NextToken member contains a token that you pass in the next call to DescribeDirectories to retrieve the next set of items.

You can also specify a maximum number of return results with the Limit parameter.

This operation returns paginated results.

Creating a Request

data DescribeDirectories

Contains the inputs for the DescribeDirectories operation.

See: newDescribeDirectories smart constructor.

Constructors

DescribeDirectories'

Fields

nextToken :: Maybe Text

The DescribeDirectoriesResult.NextToken value from a previous call to DescribeDirectories. Pass null if this is the first call.

directoryIds :: Maybe [Text]

A list of identifiers of the directories for which to obtain the information. If this member is null, all directories that belong to the current account are returned.

An empty list results in an InvalidParameterException being thrown.

limit :: Maybe Natural

The maximum number of items to return. If this value is zero, the maximum number of items is specified by the limitations of the operation.

newDescribeDirectories :: DescribeDirectories

Create a value of DescribeDirectories with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeDirectories', describeDirectories_nextToken - The DescribeDirectoriesResult.NextToken value from a previous call to DescribeDirectories. Pass null if this is the first call.

$sel:directoryIds:DescribeDirectories', describeDirectories_directoryIds - A list of identifiers of the directories for which to obtain the information. If this member is null, all directories that belong to the current account are returned.

An empty list results in an InvalidParameterException being thrown.

$sel:limit:DescribeDirectories', describeDirectories_limit - The maximum number of items to return. If this value is zero, the maximum number of items is specified by the limitations of the operation.

Request Lenses

describeDirectories_nextToken :: Lens' DescribeDirectories (Maybe Text)

The DescribeDirectoriesResult.NextToken value from a previous call to DescribeDirectories. Pass null if this is the first call.

describeDirectories_directoryIds :: Lens' DescribeDirectories (Maybe [Text])

A list of identifiers of the directories for which to obtain the information. If this member is null, all directories that belong to the current account are returned.

An empty list results in an InvalidParameterException being thrown.

describeDirectories_limit :: Lens' DescribeDirectories (Maybe Natural)

The maximum number of items to return. If this value is zero, the maximum number of items is specified by the limitations of the operation.

Destructuring the Response

data DescribeDirectoriesResponse

Contains the results of the DescribeDirectories operation.

See: newDescribeDirectoriesResponse smart constructor.

Constructors

DescribeDirectoriesResponse'

Fields

nextToken :: Maybe Text

If not null, more results are available. Pass this value for the NextToken parameter in a subsequent call to DescribeDirectories to retrieve the next set of items.

directoryDescriptions :: Maybe [DirectoryDescription]

The list of DirectoryDescription objects that were retrieved.

It is possible that this list contains less than the number of items specified in the Limit member of the request. This occurs if there are less than the requested number of items left to retrieve, or if the limitations of the operation have been exceeded.

httpStatus :: Int

The response's http status code.

newDescribeDirectoriesResponse

Arguments

:: Int
$sel:httpStatus:DescribeDirectoriesResponse'

-> DescribeDirectoriesResponse

Create a value of DescribeDirectoriesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeDirectories', describeDirectoriesResponse_nextToken - If not null, more results are available. Pass this value for the NextToken parameter in a subsequent call to DescribeDirectories to retrieve the next set of items.

$sel:directoryDescriptions:DescribeDirectoriesResponse', describeDirectoriesResponse_directoryDescriptions - The list of DirectoryDescription objects that were retrieved.

It is possible that this list contains less than the number of items specified in the Limit member of the request. This occurs if there are less than the requested number of items left to retrieve, or if the limitations of the operation have been exceeded.

$sel:httpStatus:DescribeDirectoriesResponse', describeDirectoriesResponse_httpStatus - The response's http status code.

Response Lenses

describeDirectoriesResponse_nextToken :: Lens' DescribeDirectoriesResponse (Maybe Text)

If not null, more results are available. Pass this value for the NextToken parameter in a subsequent call to DescribeDirectories to retrieve the next set of items.

describeDirectoriesResponse_directoryDescriptions :: Lens' DescribeDirectoriesResponse (Maybe [DirectoryDescription])

The list of DirectoryDescription objects that were retrieved.

It is possible that this list contains less than the number of items specified in the Limit member of the request. This occurs if there are less than the requested number of items left to retrieve, or if the limitations of the operation have been exceeded.

describeDirectoriesResponse_httpStatus :: Lens' DescribeDirectoriesResponse Int

The response's http status code.