Network.AWS.DirectoryService.DescribeDomainControllers

Description

Provides information about any domain controllers in your directory.

This operation returns paginated results.

Creating a Request

data DescribeDomainControllers

See: newDescribeDomainControllers smart constructor.

Constructors

DescribeDomainControllers'

Fields

nextToken :: Maybe Text

The DescribeDomainControllers.NextToken value from a previous call to DescribeDomainControllers. Pass null if this is the first call.

domainControllerIds :: Maybe [Text]

A list of identifiers for the domain controllers whose information will be provided.

limit :: Maybe Natural

The maximum number of items to return.

directoryId :: Text

Identifier of the directory for which to retrieve the domain controller information.

newDescribeDomainControllers

Arguments

:: Text
$sel:directoryId:DescribeDomainControllers'

-> DescribeDomainControllers

Create a value of DescribeDomainControllers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeDomainControllers', describeDomainControllers_nextToken - The DescribeDomainControllers.NextToken value from a previous call to DescribeDomainControllers. Pass null if this is the first call.

$sel:domainControllerIds:DescribeDomainControllers', describeDomainControllers_domainControllerIds - A list of identifiers for the domain controllers whose information will be provided.

$sel:limit:DescribeDomainControllers', describeDomainControllers_limit - The maximum number of items to return.

$sel:directoryId:DescribeDomainControllers', describeDomainControllers_directoryId - Identifier of the directory for which to retrieve the domain controller information.

Request Lenses

describeDomainControllers_nextToken :: Lens' DescribeDomainControllers (Maybe Text)

The DescribeDomainControllers.NextToken value from a previous call to DescribeDomainControllers. Pass null if this is the first call.

describeDomainControllers_domainControllerIds :: Lens' DescribeDomainControllers (Maybe [Text])

A list of identifiers for the domain controllers whose information will be provided.

describeDomainControllers_limit :: Lens' DescribeDomainControllers (Maybe Natural)

The maximum number of items to return.

describeDomainControllers_directoryId :: Lens' DescribeDomainControllers Text

Identifier of the directory for which to retrieve the domain controller information.

Destructuring the Response

data DescribeDomainControllersResponse

See: newDescribeDomainControllersResponse smart constructor.

Constructors

DescribeDomainControllersResponse'

Fields

nextToken :: Maybe Text

If not null, more results are available. Pass this value for the NextToken parameter in a subsequent call to DescribeDomainControllers retrieve the next set of items.

domainControllers :: Maybe [DomainController]

List of the DomainController objects that were retrieved.

httpStatus :: Int

The response's http status code.

newDescribeDomainControllersResponse

Arguments

:: Int
$sel:httpStatus:DescribeDomainControllersResponse'

-> DescribeDomainControllersResponse

Create a value of DescribeDomainControllersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeDomainControllers', describeDomainControllersResponse_nextToken - If not null, more results are available. Pass this value for the NextToken parameter in a subsequent call to DescribeDomainControllers retrieve the next set of items.

$sel:domainControllers:DescribeDomainControllersResponse', describeDomainControllersResponse_domainControllers - List of the DomainController objects that were retrieved.

$sel:httpStatus:DescribeDomainControllersResponse', describeDomainControllersResponse_httpStatus - The response's http status code.

Response Lenses

describeDomainControllersResponse_nextToken :: Lens' DescribeDomainControllersResponse (Maybe Text)

If not null, more results are available. Pass this value for the NextToken parameter in a subsequent call to DescribeDomainControllers retrieve the next set of items.

describeDomainControllersResponse_domainControllers :: Lens' DescribeDomainControllersResponse (Maybe [DomainController])

List of the DomainController objects that were retrieved.

describeDomainControllersResponse_httpStatus :: Lens' DescribeDomainControllersResponse Int

The response's http status code.