Network.AWS.DirectoryService.DescribeSharedDirectories

Description

Returns the shared directories in your account.

This operation returns paginated results.

Creating a Request

data DescribeSharedDirectories

See: newDescribeSharedDirectories smart constructor.

Constructors

DescribeSharedDirectories'

Fields

nextToken :: Maybe Text

The DescribeSharedDirectoriesResult.NextToken value from a previous call to DescribeSharedDirectories. Pass null if this is the first call.

sharedDirectoryIds :: Maybe [Text]

A list of identifiers of all shared directories in your account.

limit :: Maybe Natural

The number of shared directories to return in the response object.

ownerDirectoryId :: Text

Returns the identifier of the directory in the directory owner account.

newDescribeSharedDirectories

Arguments

:: Text
$sel:ownerDirectoryId:DescribeSharedDirectories'

-> DescribeSharedDirectories

Create a value of DescribeSharedDirectories with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeSharedDirectories', describeSharedDirectories_nextToken - The DescribeSharedDirectoriesResult.NextToken value from a previous call to DescribeSharedDirectories. Pass null if this is the first call.

$sel:sharedDirectoryIds:DescribeSharedDirectories', describeSharedDirectories_sharedDirectoryIds - A list of identifiers of all shared directories in your account.

$sel:limit:DescribeSharedDirectories', describeSharedDirectories_limit - The number of shared directories to return in the response object.

$sel:ownerDirectoryId:DescribeSharedDirectories', describeSharedDirectories_ownerDirectoryId - Returns the identifier of the directory in the directory owner account.

Request Lenses

describeSharedDirectories_nextToken :: Lens' DescribeSharedDirectories (Maybe Text)

The DescribeSharedDirectoriesResult.NextToken value from a previous call to DescribeSharedDirectories. Pass null if this is the first call.

describeSharedDirectories_sharedDirectoryIds :: Lens' DescribeSharedDirectories (Maybe [Text])

A list of identifiers of all shared directories in your account.

describeSharedDirectories_limit :: Lens' DescribeSharedDirectories (Maybe Natural)

The number of shared directories to return in the response object.

describeSharedDirectories_ownerDirectoryId :: Lens' DescribeSharedDirectories Text

Returns the identifier of the directory in the directory owner account.

Destructuring the Response

data DescribeSharedDirectoriesResponse

See: newDescribeSharedDirectoriesResponse smart constructor.

Constructors

DescribeSharedDirectoriesResponse'

Fields

nextToken :: Maybe Text

If not null, token that indicates that more results are available. Pass this value for the NextToken parameter in a subsequent call to DescribeSharedDirectories to retrieve the next set of items.

sharedDirectories :: Maybe [SharedDirectory]

A list of all shared directories in your account.

httpStatus :: Int

The response's http status code.

newDescribeSharedDirectoriesResponse

Arguments

:: Int
$sel:httpStatus:DescribeSharedDirectoriesResponse'

-> DescribeSharedDirectoriesResponse

Create a value of DescribeSharedDirectoriesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeSharedDirectories', describeSharedDirectoriesResponse_nextToken - If not null, token that indicates that more results are available. Pass this value for the NextToken parameter in a subsequent call to DescribeSharedDirectories to retrieve the next set of items.

$sel:sharedDirectories:DescribeSharedDirectoriesResponse', describeSharedDirectoriesResponse_sharedDirectories - A list of all shared directories in your account.

$sel:httpStatus:DescribeSharedDirectoriesResponse', describeSharedDirectoriesResponse_httpStatus - The response's http status code.

Response Lenses

describeSharedDirectoriesResponse_nextToken :: Lens' DescribeSharedDirectoriesResponse (Maybe Text)

If not null, token that indicates that more results are available. Pass this value for the NextToken parameter in a subsequent call to DescribeSharedDirectories to retrieve the next set of items.

describeSharedDirectoriesResponse_sharedDirectories :: Lens' DescribeSharedDirectoriesResponse (Maybe [SharedDirectory])

A list of all shared directories in your account.

describeSharedDirectoriesResponse_httpStatus :: Lens' DescribeSharedDirectoriesResponse Int

The response's http status code.