Network.AWS.DirectoryService.DescribeSnapshots

Description

Obtains information about the directory snapshots that belong to this account.

This operation supports pagination with the use of the NextToken request and response parameters. If more results are available, the DescribeSnapshots.NextToken member contains a token that you pass in the next call to DescribeSnapshots to retrieve the next set of items.

You can also specify a maximum number of return results with the Limit parameter.

This operation returns paginated results.

Creating a Request

data DescribeSnapshots

Contains the inputs for the DescribeSnapshots operation.

See: newDescribeSnapshots smart constructor.

Constructors

DescribeSnapshots'

Fields

nextToken :: Maybe Text

The DescribeSnapshotsResult.NextToken value from a previous call to DescribeSnapshots. Pass null if this is the first call.

snapshotIds :: Maybe [Text]

A list of identifiers of the snapshots to obtain the information for. If this member is null or empty, all snapshots are returned using the Limit and NextToken members.

directoryId :: Maybe Text

The identifier of the directory for which to retrieve snapshot information.

limit :: Maybe Natural

The maximum number of objects to return.

newDescribeSnapshots :: DescribeSnapshots

Create a value of DescribeSnapshots with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeSnapshots', describeSnapshots_nextToken - The DescribeSnapshotsResult.NextToken value from a previous call to DescribeSnapshots. Pass null if this is the first call.

$sel:snapshotIds:DescribeSnapshots', describeSnapshots_snapshotIds - A list of identifiers of the snapshots to obtain the information for. If this member is null or empty, all snapshots are returned using the Limit and NextToken members.

$sel:directoryId:DescribeSnapshots', describeSnapshots_directoryId - The identifier of the directory for which to retrieve snapshot information.

$sel:limit:DescribeSnapshots', describeSnapshots_limit - The maximum number of objects to return.

Request Lenses

describeSnapshots_nextToken :: Lens' DescribeSnapshots (Maybe Text)

The DescribeSnapshotsResult.NextToken value from a previous call to DescribeSnapshots. Pass null if this is the first call.

describeSnapshots_snapshotIds :: Lens' DescribeSnapshots (Maybe [Text])

A list of identifiers of the snapshots to obtain the information for. If this member is null or empty, all snapshots are returned using the Limit and NextToken members.

describeSnapshots_directoryId :: Lens' DescribeSnapshots (Maybe Text)

The identifier of the directory for which to retrieve snapshot information.

describeSnapshots_limit :: Lens' DescribeSnapshots (Maybe Natural)

The maximum number of objects to return.

Destructuring the Response

data DescribeSnapshotsResponse

Contains the results of the DescribeSnapshots operation.

See: newDescribeSnapshotsResponse smart constructor.

Constructors

DescribeSnapshotsResponse'

Fields

snapshots :: Maybe [Snapshot]

The list of Snapshot objects that were retrieved.

It is possible that this list contains less than the number of items specified in the Limit member of the request. This occurs if there are less than the requested number of items left to retrieve, or if the limitations of the operation have been exceeded.

nextToken :: Maybe Text

If not null, more results are available. Pass this value in the NextToken member of a subsequent call to DescribeSnapshots.

httpStatus :: Int

The response's http status code.

newDescribeSnapshotsResponse

Arguments

:: Int
$sel:httpStatus:DescribeSnapshotsResponse'

-> DescribeSnapshotsResponse

Create a value of DescribeSnapshotsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:snapshots:DescribeSnapshotsResponse', describeSnapshotsResponse_snapshots - The list of Snapshot objects that were retrieved.

It is possible that this list contains less than the number of items specified in the Limit member of the request. This occurs if there are less than the requested number of items left to retrieve, or if the limitations of the operation have been exceeded.

$sel:nextToken:DescribeSnapshots', describeSnapshotsResponse_nextToken - If not null, more results are available. Pass this value in the NextToken member of a subsequent call to DescribeSnapshots.

$sel:httpStatus:DescribeSnapshotsResponse', describeSnapshotsResponse_httpStatus - The response's http status code.

Response Lenses

describeSnapshotsResponse_snapshots :: Lens' DescribeSnapshotsResponse (Maybe [Snapshot])

The list of Snapshot objects that were retrieved.

It is possible that this list contains less than the number of items specified in the Limit member of the request. This occurs if there are less than the requested number of items left to retrieve, or if the limitations of the operation have been exceeded.

describeSnapshotsResponse_nextToken :: Lens' DescribeSnapshotsResponse (Maybe Text)

If not null, more results are available. Pass this value in the NextToken member of a subsequent call to DescribeSnapshots.

describeSnapshotsResponse_httpStatus :: Lens' DescribeSnapshotsResponse Int

The response's http status code.