Network.AWS.DirectoryService.EnableSso

Description

Enables single sign-on for a directory. Single sign-on allows users in your directory to access certain Amazon Web Services services from a computer joined to the directory without having to enter their credentials separately.

Creating a Request

data EnableSso

Contains the inputs for the EnableSso operation.

See: newEnableSso smart constructor.

Constructors

EnableSso'

Fields

password :: Maybe (Sensitive Text)

The password of an alternate account to use to enable single-sign on. This is only used for AD Connector directories. For more information, see the UserName parameter.

userName :: Maybe Text

The username of an alternate account to use to enable single-sign on. This is only used for AD Connector directories. This account must have privileges to add a service principal name.

If the AD Connector service account does not have privileges to add a service principal name, you can specify an alternate account with the UserName and Password parameters. These credentials are only used to enable single sign-on and are not stored by the service. The AD Connector service account is not changed.

directoryId :: Text

The identifier of the directory for which to enable single-sign on.

newEnableSso

Arguments

:: Text
$sel:directoryId:EnableSso'

-> EnableSso

Create a value of EnableSso with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:password:EnableSso', enableSso_password - The password of an alternate account to use to enable single-sign on. This is only used for AD Connector directories. For more information, see the UserName parameter.

$sel:userName:EnableSso', enableSso_userName - The username of an alternate account to use to enable single-sign on. This is only used for AD Connector directories. This account must have privileges to add a service principal name.

If the AD Connector service account does not have privileges to add a service principal name, you can specify an alternate account with the UserName and Password parameters. These credentials are only used to enable single sign-on and are not stored by the service. The AD Connector service account is not changed.

$sel:directoryId:EnableSso', enableSso_directoryId - The identifier of the directory for which to enable single-sign on.

Request Lenses

enableSso_password :: Lens' EnableSso (Maybe Text)

The password of an alternate account to use to enable single-sign on. This is only used for AD Connector directories. For more information, see the UserName parameter.

enableSso_userName :: Lens' EnableSso (Maybe Text)

The username of an alternate account to use to enable single-sign on. This is only used for AD Connector directories. This account must have privileges to add a service principal name.

If the AD Connector service account does not have privileges to add a service principal name, you can specify an alternate account with the UserName and Password parameters. These credentials are only used to enable single sign-on and are not stored by the service. The AD Connector service account is not changed.

enableSso_directoryId :: Lens' EnableSso Text

The identifier of the directory for which to enable single-sign on.

Destructuring the Response

data EnableSsoResponse

Contains the results of the EnableSso operation.

See: newEnableSsoResponse smart constructor.

Constructors

EnableSsoResponse'

Fields

httpStatus :: Int

The response's http status code.

newEnableSsoResponse

Arguments

:: Int
$sel:httpStatus:EnableSsoResponse'

-> EnableSsoResponse

Create a value of EnableSsoResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:EnableSsoResponse', enableSsoResponse_httpStatus - The response's http status code.

Response Lenses

enableSsoResponse_httpStatus :: Lens' EnableSsoResponse Int

The response's http status code.