Network.AWS.DirectoryService.GetSnapshotLimits

Description

Obtains the manual snapshot limits for a directory.

Creating a Request

data GetSnapshotLimits

Contains the inputs for the GetSnapshotLimits operation.

See: newGetSnapshotLimits smart constructor.

Constructors

GetSnapshotLimits'

Fields

directoryId :: Text

Contains the identifier of the directory to obtain the limits for.

newGetSnapshotLimits

Arguments

:: Text
$sel:directoryId:GetSnapshotLimits'

-> GetSnapshotLimits

Create a value of GetSnapshotLimits with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:directoryId:GetSnapshotLimits', getSnapshotLimits_directoryId - Contains the identifier of the directory to obtain the limits for.

Request Lenses

getSnapshotLimits_directoryId :: Lens' GetSnapshotLimits Text

Contains the identifier of the directory to obtain the limits for.

Destructuring the Response

data GetSnapshotLimitsResponse

Contains the results of the GetSnapshotLimits operation.

See: newGetSnapshotLimitsResponse smart constructor.

Constructors

GetSnapshotLimitsResponse'

Fields

snapshotLimits :: Maybe SnapshotLimits

A SnapshotLimits object that contains the manual snapshot limits for the specified directory.

httpStatus :: Int

The response's http status code.

newGetSnapshotLimitsResponse

Arguments

:: Int
$sel:httpStatus:GetSnapshotLimitsResponse'

-> GetSnapshotLimitsResponse

Create a value of GetSnapshotLimitsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:snapshotLimits:GetSnapshotLimitsResponse', getSnapshotLimitsResponse_snapshotLimits - A SnapshotLimits object that contains the manual snapshot limits for the specified directory.

$sel:httpStatus:GetSnapshotLimitsResponse', getSnapshotLimitsResponse_httpStatus - The response's http status code.

Response Lenses

getSnapshotLimitsResponse_snapshotLimits :: Lens' GetSnapshotLimitsResponse (Maybe SnapshotLimits)

A SnapshotLimits object that contains the manual snapshot limits for the specified directory.

getSnapshotLimitsResponse_httpStatus :: Lens' GetSnapshotLimitsResponse Int

The response's http status code.