Network.AWS.DirectoryService.RegisterCertificate

Description

Registers a certificate for a secure LDAP or client certificate authentication.

Creating a Request

data RegisterCertificate

See: newRegisterCertificate smart constructor.

Constructors

RegisterCertificate'

Fields

clientCertAuthSettings :: Maybe ClientCertAuthSettings

A ClientCertAuthSettings object that contains client certificate authentication settings.

type' :: Maybe CertificateType

The function that the registered certificate performs. Valid values include ClientLDAPS or ClientCertAuth. The default value is ClientLDAPS.

directoryId :: Text

The identifier of the directory.

certificateData :: Text

The certificate PEM string that needs to be registered.

newRegisterCertificate

Arguments

:: Text
$sel:directoryId:RegisterCertificate'

-> Text
$sel:certificateData:RegisterCertificate'

-> RegisterCertificate

Create a value of RegisterCertificate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientCertAuthSettings:RegisterCertificate', registerCertificate_clientCertAuthSettings - A ClientCertAuthSettings object that contains client certificate authentication settings.

$sel:type':RegisterCertificate', registerCertificate_type - The function that the registered certificate performs. Valid values include ClientLDAPS or ClientCertAuth. The default value is ClientLDAPS.

$sel:directoryId:RegisterCertificate', registerCertificate_directoryId - The identifier of the directory.

$sel:certificateData:RegisterCertificate', registerCertificate_certificateData - The certificate PEM string that needs to be registered.

Request Lenses

registerCertificate_clientCertAuthSettings :: Lens' RegisterCertificate (Maybe ClientCertAuthSettings)

A ClientCertAuthSettings object that contains client certificate authentication settings.

registerCertificate_type :: Lens' RegisterCertificate (Maybe CertificateType)

The function that the registered certificate performs. Valid values include ClientLDAPS or ClientCertAuth. The default value is ClientLDAPS.

registerCertificate_directoryId :: Lens' RegisterCertificate Text

The identifier of the directory.

registerCertificate_certificateData :: Lens' RegisterCertificate Text

The certificate PEM string that needs to be registered.

Destructuring the Response

data RegisterCertificateResponse

See: newRegisterCertificateResponse smart constructor.

Constructors

RegisterCertificateResponse'

Fields

certificateId :: Maybe Text

The identifier of the certificate.

httpStatus :: Int

The response's http status code.

newRegisterCertificateResponse

Arguments

:: Int
$sel:httpStatus:RegisterCertificateResponse'

-> RegisterCertificateResponse

Create a value of RegisterCertificateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:certificateId:RegisterCertificateResponse', registerCertificateResponse_certificateId - The identifier of the certificate.

$sel:httpStatus:RegisterCertificateResponse', registerCertificateResponse_httpStatus - The response's http status code.

Response Lenses

registerCertificateResponse_certificateId :: Lens' RegisterCertificateResponse (Maybe Text)

The identifier of the certificate.

registerCertificateResponse_httpStatus :: Lens' RegisterCertificateResponse Int

The response's http status code.