Network.AWS.DirectoryService.RegisterEventTopic

Description

Associates a directory with an Amazon SNS topic. This establishes the directory as a publisher to the specified Amazon SNS topic. You can then receive email or text (SMS) messages when the status of your directory changes. You get notified if your directory goes from an Active status to an Impaired or Inoperable status. You also receive a notification when the directory returns to an Active status.

Creating a Request

data RegisterEventTopic

Registers a new event topic.

See: newRegisterEventTopic smart constructor.

Constructors

RegisterEventTopic'

Fields

directoryId :: Text

The Directory ID that will publish status messages to the Amazon SNS topic.

topicName :: Text

The Amazon SNS topic name to which the directory will publish status messages. This Amazon SNS topic must be in the same region as the specified Directory ID.

newRegisterEventTopic

Arguments

:: Text
$sel:directoryId:RegisterEventTopic'

-> Text
$sel:topicName:RegisterEventTopic'

-> RegisterEventTopic

Create a value of RegisterEventTopic with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:directoryId:RegisterEventTopic', registerEventTopic_directoryId - The Directory ID that will publish status messages to the Amazon SNS topic.

$sel:topicName:RegisterEventTopic', registerEventTopic_topicName - The Amazon SNS topic name to which the directory will publish status messages. This Amazon SNS topic must be in the same region as the specified Directory ID.

Request Lenses

registerEventTopic_directoryId :: Lens' RegisterEventTopic Text

The Directory ID that will publish status messages to the Amazon SNS topic.

registerEventTopic_topicName :: Lens' RegisterEventTopic Text

The Amazon SNS topic name to which the directory will publish status messages. This Amazon SNS topic must be in the same region as the specified Directory ID.

Destructuring the Response

data RegisterEventTopicResponse

The result of a RegisterEventTopic request.

See: newRegisterEventTopicResponse smart constructor.

Constructors

RegisterEventTopicResponse'

Fields

httpStatus :: Int

The response's http status code.

newRegisterEventTopicResponse

Arguments

:: Int
$sel:httpStatus:RegisterEventTopicResponse'

-> RegisterEventTopicResponse

Create a value of RegisterEventTopicResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:RegisterEventTopicResponse', registerEventTopicResponse_httpStatus - The response's http status code.

Response Lenses

registerEventTopicResponse_httpStatus :: Lens' RegisterEventTopicResponse Int

The response's http status code.