Network.AWS.DirectoryService.RemoveIpRoutes

Description

Removes IP address blocks from a directory.

Creating a Request

data RemoveIpRoutes

See: newRemoveIpRoutes smart constructor.

Constructors

RemoveIpRoutes'

Fields

directoryId :: Text

Identifier (ID) of the directory from which you want to remove the IP addresses.

cidrIps :: [Text]

IP address blocks that you want to remove.

newRemoveIpRoutes

Arguments

:: Text
$sel:directoryId:RemoveIpRoutes'

-> RemoveIpRoutes

Create a value of RemoveIpRoutes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:directoryId:RemoveIpRoutes', removeIpRoutes_directoryId - Identifier (ID) of the directory from which you want to remove the IP addresses.

$sel:cidrIps:RemoveIpRoutes', removeIpRoutes_cidrIps - IP address blocks that you want to remove.

Request Lenses

removeIpRoutes_directoryId :: Lens' RemoveIpRoutes Text

Identifier (ID) of the directory from which you want to remove the IP addresses.

removeIpRoutes_cidrIps :: Lens' RemoveIpRoutes [Text]

IP address blocks that you want to remove.

Destructuring the Response

data RemoveIpRoutesResponse

See: newRemoveIpRoutesResponse smart constructor.

Constructors

RemoveIpRoutesResponse'

Fields

httpStatus :: Int

The response's http status code.

newRemoveIpRoutesResponse

Arguments

:: Int
$sel:httpStatus:RemoveIpRoutesResponse'

-> RemoveIpRoutesResponse

Create a value of RemoveIpRoutesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:RemoveIpRoutesResponse', removeIpRoutesResponse_httpStatus - The response's http status code.

Response Lenses

removeIpRoutesResponse_httpStatus :: Lens' RemoveIpRoutesResponse Int

The response's http status code.