Network.AWS.DirectoryService.ResetUserPassword

Description

Resets the password for any user in your Managed Microsoft AD or Simple AD directory.

You can reset the password for any user in your directory with the following exceptions:

For Simple AD, you cannot reset the password for any user that is a member of either the Domain Admins or Enterprise Admins group except for the administrator user.

For Managed Microsoft AD, you can only reset the password for a user that is in an OU based off of the NetBIOS name that you typed when you created your directory. For example, you cannot reset the password for a user in the Amazon Web Services Reserved OU. For more information about the OU structure for an Managed Microsoft AD directory, see What Gets Created in the Directory Service Administration Guide.

Creating a Request

data ResetUserPassword

See: newResetUserPassword smart constructor.

Constructors

ResetUserPassword'

Fields

directoryId :: Text

Identifier of the Managed Microsoft AD or Simple AD directory in which the user resides.

userName :: Text

The user name of the user whose password will be reset.

newPassword' :: Sensitive Text

The new password that will be reset.

newResetUserPassword

Arguments

:: Text
$sel:directoryId:ResetUserPassword'

-> Text
$sel:userName:ResetUserPassword'

-> Text
$sel:newPassword':ResetUserPassword'

-> ResetUserPassword

Create a value of ResetUserPassword with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:directoryId:ResetUserPassword', resetUserPassword_directoryId - Identifier of the Managed Microsoft AD or Simple AD directory in which the user resides.

$sel:userName:ResetUserPassword', resetUserPassword_userName - The user name of the user whose password will be reset.

$sel:newPassword':ResetUserPassword', resetUserPassword_newPassword - The new password that will be reset.

Request Lenses

resetUserPassword_directoryId :: Lens' ResetUserPassword Text

Identifier of the Managed Microsoft AD or Simple AD directory in which the user resides.

resetUserPassword_userName :: Lens' ResetUserPassword Text

The user name of the user whose password will be reset.

resetUserPassword_newPassword :: Lens' ResetUserPassword Text

The new password that will be reset.

Destructuring the Response

data ResetUserPasswordResponse

See: newResetUserPasswordResponse smart constructor.

Constructors

ResetUserPasswordResponse'

Fields

httpStatus :: Int

The response's http status code.

newResetUserPasswordResponse

Arguments

:: Int
$sel:httpStatus:ResetUserPasswordResponse'

-> ResetUserPasswordResponse

Create a value of ResetUserPasswordResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:ResetUserPasswordResponse', resetUserPasswordResponse_httpStatus - The response's http status code.

Response Lenses

resetUserPasswordResponse_httpStatus :: Lens' ResetUserPasswordResponse Int

The response's http status code.