Network.AWS.DirectoryService.Types.DirectoryConnectSettingsDescription

Description

Documentation

data DirectoryConnectSettingsDescription

Contains information about an AD Connector directory.

See: newDirectoryConnectSettingsDescription smart constructor.

Constructors

DirectoryConnectSettingsDescription'

Fields

availabilityZones :: Maybe [Text]

A list of the Availability Zones that the directory is in.

securityGroupId :: Maybe Text

The security group identifier for the AD Connector directory.

subnetIds :: Maybe [Text]

A list of subnet identifiers in the VPC that the AD Connector is in.

customerUserName :: Maybe Text

The user name of the service account in your self-managed directory.

connectIps :: Maybe [Text]

The IP addresses of the AD Connector servers.

vpcId :: Maybe Text

The identifier of the VPC that the AD Connector is in.

newDirectoryConnectSettingsDescription :: DirectoryConnectSettingsDescription

Create a value of DirectoryConnectSettingsDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:availabilityZones:DirectoryConnectSettingsDescription', directoryConnectSettingsDescription_availabilityZones - A list of the Availability Zones that the directory is in.

$sel:securityGroupId:DirectoryConnectSettingsDescription', directoryConnectSettingsDescription_securityGroupId - The security group identifier for the AD Connector directory.

$sel:subnetIds:DirectoryConnectSettingsDescription', directoryConnectSettingsDescription_subnetIds - A list of subnet identifiers in the VPC that the AD Connector is in.

$sel:customerUserName:DirectoryConnectSettingsDescription', directoryConnectSettingsDescription_customerUserName - The user name of the service account in your self-managed directory.

$sel:connectIps:DirectoryConnectSettingsDescription', directoryConnectSettingsDescription_connectIps - The IP addresses of the AD Connector servers.

$sel:vpcId:DirectoryConnectSettingsDescription', directoryConnectSettingsDescription_vpcId - The identifier of the VPC that the AD Connector is in.

directoryConnectSettingsDescription_availabilityZones :: Lens' DirectoryConnectSettingsDescription (Maybe [Text])

A list of the Availability Zones that the directory is in.

directoryConnectSettingsDescription_securityGroupId :: Lens' DirectoryConnectSettingsDescription (Maybe Text)

The security group identifier for the AD Connector directory.

directoryConnectSettingsDescription_subnetIds :: Lens' DirectoryConnectSettingsDescription (Maybe [Text])

A list of subnet identifiers in the VPC that the AD Connector is in.

directoryConnectSettingsDescription_customerUserName :: Lens' DirectoryConnectSettingsDescription (Maybe Text)

The user name of the service account in your self-managed directory.

directoryConnectSettingsDescription_connectIps :: Lens' DirectoryConnectSettingsDescription (Maybe [Text])

The IP addresses of the AD Connector servers.

directoryConnectSettingsDescription_vpcId :: Lens' DirectoryConnectSettingsDescription (Maybe Text)

The identifier of the VPC that the AD Connector is in.