Network.AWS.DirectoryService.Types.DirectoryDescription

Description

Documentation

data DirectoryDescription

Contains information about an Directory Service directory.

See: newDirectoryDescription smart constructor.

Constructors

DirectoryDescription'

Fields

radiusStatus :: Maybe RadiusStatus

The status of the RADIUS MFA server connection.

alias :: Maybe Text

The alias for the directory. If no alias has been created for the directory, the alias is the directory identifier, such as d-XXXXXXXXXX.

shareNotes :: Maybe (Sensitive Text)

A directory share request that is sent by the directory owner to the directory consumer. The request includes a typed message to help the directory consumer administrator determine whether to approve or reject the share invitation.

connectSettings :: Maybe DirectoryConnectSettingsDescription

A DirectoryConnectSettingsDescription object that contains additional information about an AD Connector directory. This member is only present if the directory is an AD Connector directory.

stageReason :: Maybe Text

Additional information about the directory stage.

launchTime :: Maybe POSIX

Specifies when the directory was created.

vpcSettings :: Maybe DirectoryVpcSettingsDescription

A DirectoryVpcSettingsDescription object that contains additional information about a directory. This member is only present if the directory is a Simple AD or Managed Microsoft AD directory.

regionsInfo :: Maybe RegionsInfo

Lists the Regions where the directory has replicated.

shortName :: Maybe Text

The short name of the directory.

shareMethod :: Maybe ShareMethod

The method used when sharing a directory to determine whether the directory should be shared within your Amazon Web Services organization (ORGANIZATIONS) or with any Amazon Web Services account by sending a shared directory request (HANDSHAKE).

name :: Maybe Text

The fully qualified name of the directory.

accessUrl :: Maybe Text

The access URL for the directory, such as http://<alias>.awsapps.com. If no alias has been created for the directory, <alias> is the directory identifier, such as d-XXXXXXXXXX.

directoryId :: Maybe Text

The directory identifier.

stage :: Maybe DirectoryStage

The current stage of the directory.

edition :: Maybe DirectoryEdition

The edition associated with this directory.

shareStatus :: Maybe ShareStatus

Current directory status of the shared Managed Microsoft AD directory.

description :: Maybe Text

The description for the directory.

ownerDirectoryDescription :: Maybe OwnerDirectoryDescription

Describes the Managed Microsoft AD directory in the directory owner account.

type' :: Maybe DirectoryType

The directory size.

dnsIpAddrs :: Maybe [Text]

The IP addresses of the DNS servers for the directory. For a Simple AD or Microsoft AD directory, these are the IP addresses of the Simple AD or Microsoft AD directory servers. For an AD Connector directory, these are the IP addresses of the DNS servers or domain controllers in your self-managed directory to which the AD Connector is connected.

desiredNumberOfDomainControllers :: Maybe Natural

The desired number of domain controllers in the directory if the directory is Microsoft AD.

size :: Maybe DirectorySize

The directory size.

radiusSettings :: Maybe RadiusSettings

A RadiusSettings object that contains information about the RADIUS server configured for this directory.

stageLastUpdatedDateTime :: Maybe POSIX

The date and time that the stage was last updated.

ssoEnabled :: Maybe Bool

Indicates if single sign-on is enabled for the directory. For more information, see EnableSso and DisableSso.

newDirectoryDescription :: DirectoryDescription

Create a value of DirectoryDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:radiusStatus:DirectoryDescription', directoryDescription_radiusStatus - The status of the RADIUS MFA server connection.

$sel:alias:DirectoryDescription', directoryDescription_alias - The alias for the directory. If no alias has been created for the directory, the alias is the directory identifier, such as d-XXXXXXXXXX.

$sel:shareNotes:DirectoryDescription', directoryDescription_shareNotes - A directory share request that is sent by the directory owner to the directory consumer. The request includes a typed message to help the directory consumer administrator determine whether to approve or reject the share invitation.

$sel:connectSettings:DirectoryDescription', directoryDescription_connectSettings - A DirectoryConnectSettingsDescription object that contains additional information about an AD Connector directory. This member is only present if the directory is an AD Connector directory.

$sel:stageReason:DirectoryDescription', directoryDescription_stageReason - Additional information about the directory stage.

$sel:launchTime:DirectoryDescription', directoryDescription_launchTime - Specifies when the directory was created.

$sel:vpcSettings:DirectoryDescription', directoryDescription_vpcSettings - A DirectoryVpcSettingsDescription object that contains additional information about a directory. This member is only present if the directory is a Simple AD or Managed Microsoft AD directory.

$sel:regionsInfo:DirectoryDescription', directoryDescription_regionsInfo - Lists the Regions where the directory has replicated.

$sel:shortName:DirectoryDescription', directoryDescription_shortName - The short name of the directory.

$sel:shareMethod:DirectoryDescription', directoryDescription_shareMethod - The method used when sharing a directory to determine whether the directory should be shared within your Amazon Web Services organization (ORGANIZATIONS) or with any Amazon Web Services account by sending a shared directory request (HANDSHAKE).

$sel:name:DirectoryDescription', directoryDescription_name - The fully qualified name of the directory.

$sel:accessUrl:DirectoryDescription', directoryDescription_accessUrl - The access URL for the directory, such as http://<alias>.awsapps.com. If no alias has been created for the directory, <alias> is the directory identifier, such as d-XXXXXXXXXX.

$sel:directoryId:DirectoryDescription', directoryDescription_directoryId - The directory identifier.

$sel:stage:DirectoryDescription', directoryDescription_stage - The current stage of the directory.

$sel:edition:DirectoryDescription', directoryDescription_edition - The edition associated with this directory.

$sel:shareStatus:DirectoryDescription', directoryDescription_shareStatus - Current directory status of the shared Managed Microsoft AD directory.

$sel:description:DirectoryDescription', directoryDescription_description - The description for the directory.

$sel:ownerDirectoryDescription:DirectoryDescription', directoryDescription_ownerDirectoryDescription - Describes the Managed Microsoft AD directory in the directory owner account.

$sel:type':DirectoryDescription', directoryDescription_type - The directory size.

$sel:dnsIpAddrs:DirectoryDescription', directoryDescription_dnsIpAddrs - The IP addresses of the DNS servers for the directory. For a Simple AD or Microsoft AD directory, these are the IP addresses of the Simple AD or Microsoft AD directory servers. For an AD Connector directory, these are the IP addresses of the DNS servers or domain controllers in your self-managed directory to which the AD Connector is connected.

$sel:desiredNumberOfDomainControllers:DirectoryDescription', directoryDescription_desiredNumberOfDomainControllers - The desired number of domain controllers in the directory if the directory is Microsoft AD.

$sel:size:DirectoryDescription', directoryDescription_size - The directory size.

$sel:radiusSettings:DirectoryDescription', directoryDescription_radiusSettings - A RadiusSettings object that contains information about the RADIUS server configured for this directory.

$sel:stageLastUpdatedDateTime:DirectoryDescription', directoryDescription_stageLastUpdatedDateTime - The date and time that the stage was last updated.

$sel:ssoEnabled:DirectoryDescription', directoryDescription_ssoEnabled - Indicates if single sign-on is enabled for the directory. For more information, see EnableSso and DisableSso.

directoryDescription_radiusStatus :: Lens' DirectoryDescription (Maybe RadiusStatus)

The status of the RADIUS MFA server connection.

directoryDescription_alias :: Lens' DirectoryDescription (Maybe Text)

The alias for the directory. If no alias has been created for the directory, the alias is the directory identifier, such as d-XXXXXXXXXX.

directoryDescription_shareNotes :: Lens' DirectoryDescription (Maybe Text)

A directory share request that is sent by the directory owner to the directory consumer. The request includes a typed message to help the directory consumer administrator determine whether to approve or reject the share invitation.

directoryDescription_connectSettings :: Lens' DirectoryDescription (Maybe DirectoryConnectSettingsDescription)

A DirectoryConnectSettingsDescription object that contains additional information about an AD Connector directory. This member is only present if the directory is an AD Connector directory.

directoryDescription_stageReason :: Lens' DirectoryDescription (Maybe Text)

Additional information about the directory stage.

directoryDescription_launchTime :: Lens' DirectoryDescription (Maybe UTCTime)

Specifies when the directory was created.

directoryDescription_vpcSettings :: Lens' DirectoryDescription (Maybe DirectoryVpcSettingsDescription)

A DirectoryVpcSettingsDescription object that contains additional information about a directory. This member is only present if the directory is a Simple AD or Managed Microsoft AD directory.

directoryDescription_regionsInfo :: Lens' DirectoryDescription (Maybe RegionsInfo)

Lists the Regions where the directory has replicated.

directoryDescription_shortName :: Lens' DirectoryDescription (Maybe Text)

The short name of the directory.

directoryDescription_shareMethod :: Lens' DirectoryDescription (Maybe ShareMethod)

The method used when sharing a directory to determine whether the directory should be shared within your Amazon Web Services organization (ORGANIZATIONS) or with any Amazon Web Services account by sending a shared directory request (HANDSHAKE).

directoryDescription_name :: Lens' DirectoryDescription (Maybe Text)

The fully qualified name of the directory.

directoryDescription_accessUrl :: Lens' DirectoryDescription (Maybe Text)

The access URL for the directory, such as http://<alias>.awsapps.com. If no alias has been created for the directory, <alias> is the directory identifier, such as d-XXXXXXXXXX.

directoryDescription_directoryId :: Lens' DirectoryDescription (Maybe Text)

The directory identifier.

directoryDescription_stage :: Lens' DirectoryDescription (Maybe DirectoryStage)

The current stage of the directory.

directoryDescription_edition :: Lens' DirectoryDescription (Maybe DirectoryEdition)

The edition associated with this directory.

directoryDescription_shareStatus :: Lens' DirectoryDescription (Maybe ShareStatus)

Current directory status of the shared Managed Microsoft AD directory.

directoryDescription_description :: Lens' DirectoryDescription (Maybe Text)

The description for the directory.

directoryDescription_ownerDirectoryDescription :: Lens' DirectoryDescription (Maybe OwnerDirectoryDescription)

Describes the Managed Microsoft AD directory in the directory owner account.

directoryDescription_type :: Lens' DirectoryDescription (Maybe DirectoryType)

The directory size.

directoryDescription_dnsIpAddrs :: Lens' DirectoryDescription (Maybe [Text])

The IP addresses of the DNS servers for the directory. For a Simple AD or Microsoft AD directory, these are the IP addresses of the Simple AD or Microsoft AD directory servers. For an AD Connector directory, these are the IP addresses of the DNS servers or domain controllers in your self-managed directory to which the AD Connector is connected.

directoryDescription_desiredNumberOfDomainControllers :: Lens' DirectoryDescription (Maybe Natural)

The desired number of domain controllers in the directory if the directory is Microsoft AD.

directoryDescription_size :: Lens' DirectoryDescription (Maybe DirectorySize)

The directory size.

directoryDescription_radiusSettings :: Lens' DirectoryDescription (Maybe RadiusSettings)

A RadiusSettings object that contains information about the RADIUS server configured for this directory.

directoryDescription_stageLastUpdatedDateTime :: Lens' DirectoryDescription (Maybe UTCTime)

The date and time that the stage was last updated.

directoryDescription_ssoEnabled :: Lens' DirectoryDescription (Maybe Bool)

Indicates if single sign-on is enabled for the directory. For more information, see EnableSso and DisableSso.