Network.AWS.DirectoryService.Types.EventTopic

Description

Documentation

data EventTopic

Information about Amazon SNS topic and Directory Service directory associations.

See: newEventTopic smart constructor.

Constructors

EventTopic'

Fields

status :: Maybe TopicStatus

The topic registration status.

createdDateTime :: Maybe POSIX

The date and time of when you associated your directory with the Amazon SNS topic.

topicName :: Maybe Text

The name of an Amazon SNS topic the receives status messages from the directory.

directoryId :: Maybe Text

The Directory ID of an Directory Service directory that will publish status messages to an Amazon SNS topic.

topicArn :: Maybe Text

The Amazon SNS topic ARN (Amazon Resource Name).

newEventTopic :: EventTopic

Create a value of EventTopic with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:EventTopic', eventTopic_status - The topic registration status.

$sel:createdDateTime:EventTopic', eventTopic_createdDateTime - The date and time of when you associated your directory with the Amazon SNS topic.

$sel:topicName:EventTopic', eventTopic_topicName - The name of an Amazon SNS topic the receives status messages from the directory.

$sel:directoryId:EventTopic', eventTopic_directoryId - The Directory ID of an Directory Service directory that will publish status messages to an Amazon SNS topic.

$sel:topicArn:EventTopic', eventTopic_topicArn - The Amazon SNS topic ARN (Amazon Resource Name).

eventTopic_status :: Lens' EventTopic (Maybe TopicStatus)

The topic registration status.

eventTopic_createdDateTime :: Lens' EventTopic (Maybe UTCTime)

The date and time of when you associated your directory with the Amazon SNS topic.

eventTopic_topicName :: Lens' EventTopic (Maybe Text)

The name of an Amazon SNS topic the receives status messages from the directory.

eventTopic_directoryId :: Lens' EventTopic (Maybe Text)

The Directory ID of an Directory Service directory that will publish status messages to an Amazon SNS topic.

eventTopic_topicArn :: Lens' EventTopic (Maybe Text)

The Amazon SNS topic ARN (Amazon Resource Name).