Network.AWS.DirectoryService.Types.IpRouteInfo

Description

Documentation

data IpRouteInfo

Information about one or more IP address blocks.

See: newIpRouteInfo smart constructor.

Constructors

IpRouteInfo'

Fields

directoryId :: Maybe Text

Identifier (ID) of the directory associated with the IP addresses.

ipRouteStatusReason :: Maybe Text

The reason for the IpRouteStatusMsg.

addedDateTime :: Maybe POSIX

The date and time the address block was added to the directory.

cidrIp :: Maybe Text

IP address block in the IpRoute.

ipRouteStatusMsg :: Maybe IpRouteStatusMsg

The status of the IP address block.

description :: Maybe Text

Description of the IpRouteInfo.

newIpRouteInfo :: IpRouteInfo

Create a value of IpRouteInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:directoryId:IpRouteInfo', ipRouteInfo_directoryId - Identifier (ID) of the directory associated with the IP addresses.

$sel:ipRouteStatusReason:IpRouteInfo', ipRouteInfo_ipRouteStatusReason - The reason for the IpRouteStatusMsg.

$sel:addedDateTime:IpRouteInfo', ipRouteInfo_addedDateTime - The date and time the address block was added to the directory.

$sel:cidrIp:IpRouteInfo', ipRouteInfo_cidrIp - IP address block in the IpRoute.

$sel:ipRouteStatusMsg:IpRouteInfo', ipRouteInfo_ipRouteStatusMsg - The status of the IP address block.

$sel:description:IpRouteInfo', ipRouteInfo_description - Description of the IpRouteInfo.

ipRouteInfo_directoryId :: Lens' IpRouteInfo (Maybe Text)

Identifier (ID) of the directory associated with the IP addresses.

ipRouteInfo_ipRouteStatusReason :: Lens' IpRouteInfo (Maybe Text)

The reason for the IpRouteStatusMsg.

ipRouteInfo_addedDateTime :: Lens' IpRouteInfo (Maybe UTCTime)

The date and time the address block was added to the directory.

ipRouteInfo_cidrIp :: Lens' IpRouteInfo (Maybe Text)

IP address block in the IpRoute.

ipRouteInfo_ipRouteStatusMsg :: Lens' IpRouteInfo (Maybe IpRouteStatusMsg)

The status of the IP address block.

ipRouteInfo_description :: Lens' IpRouteInfo (Maybe Text)

Description of the IpRouteInfo.