Network.AWS.DirectoryService.Types.OwnerDirectoryDescription

Description

Documentation

data OwnerDirectoryDescription

Describes the directory owner account details that have been shared to the directory consumer account.

See: newOwnerDirectoryDescription smart constructor.

Constructors

OwnerDirectoryDescription'

Fields

radiusStatus :: Maybe RadiusStatus

Information about the status of the RADIUS server.

accountId :: Maybe Text

Identifier of the directory owner account.

vpcSettings :: Maybe DirectoryVpcSettingsDescription

Information about the VPC settings for the directory.

directoryId :: Maybe Text

Identifier of the Managed Microsoft AD directory in the directory owner account.

dnsIpAddrs :: Maybe [Text]

IP address of the directory’s domain controllers.

radiusSettings :: Maybe RadiusSettings

A RadiusSettings object that contains information about the RADIUS server.

newOwnerDirectoryDescription :: OwnerDirectoryDescription

Create a value of OwnerDirectoryDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:radiusStatus:OwnerDirectoryDescription', ownerDirectoryDescription_radiusStatus - Information about the status of the RADIUS server.

$sel:accountId:OwnerDirectoryDescription', ownerDirectoryDescription_accountId - Identifier of the directory owner account.

$sel:vpcSettings:OwnerDirectoryDescription', ownerDirectoryDescription_vpcSettings - Information about the VPC settings for the directory.

$sel:directoryId:OwnerDirectoryDescription', ownerDirectoryDescription_directoryId - Identifier of the Managed Microsoft AD directory in the directory owner account.

$sel:dnsIpAddrs:OwnerDirectoryDescription', ownerDirectoryDescription_dnsIpAddrs - IP address of the directory’s domain controllers.

$sel:radiusSettings:OwnerDirectoryDescription', ownerDirectoryDescription_radiusSettings - A RadiusSettings object that contains information about the RADIUS server.

ownerDirectoryDescription_radiusStatus :: Lens' OwnerDirectoryDescription (Maybe RadiusStatus)

Information about the status of the RADIUS server.

ownerDirectoryDescription_accountId :: Lens' OwnerDirectoryDescription (Maybe Text)

Identifier of the directory owner account.

ownerDirectoryDescription_vpcSettings :: Lens' OwnerDirectoryDescription (Maybe DirectoryVpcSettingsDescription)

Information about the VPC settings for the directory.

ownerDirectoryDescription_directoryId :: Lens' OwnerDirectoryDescription (Maybe Text)

Identifier of the Managed Microsoft AD directory in the directory owner account.

ownerDirectoryDescription_dnsIpAddrs :: Lens' OwnerDirectoryDescription (Maybe [Text])

IP address of the directory’s domain controllers.

ownerDirectoryDescription_radiusSettings :: Lens' OwnerDirectoryDescription (Maybe RadiusSettings)

A RadiusSettings object that contains information about the RADIUS server.