Network.AWS.DirectoryService.Types.RadiusSettings

Description

Documentation

data RadiusSettings

Contains information about a Remote Authentication Dial In User Service (RADIUS) server.

See: newRadiusSettings smart constructor.

Constructors

RadiusSettings'

Fields

displayLabel :: Maybe Text

Not currently used.

radiusRetries :: Maybe Natural

The maximum number of times that communication with the RADIUS server is attempted.

authenticationProtocol :: Maybe RadiusAuthenticationProtocol

The protocol specified for your RADIUS endpoints.

radiusServers :: Maybe [Text]

An array of strings that contains the fully qualified domain name (FQDN) or IP addresses of the RADIUS server endpoints, or the FQDN or IP addresses of your RADIUS server load balancer.

useSameUsername :: Maybe Bool

Not currently used.

sharedSecret :: Maybe (Sensitive Text)

Required for enabling RADIUS on the directory.

radiusTimeout :: Maybe Natural

The amount of time, in seconds, to wait for the RADIUS server to respond.

radiusPort :: Maybe Natural

The port that your RADIUS server is using for communications. Your self-managed network must allow inbound traffic over this port from the Directory Service servers.

newRadiusSettings :: RadiusSettings

Create a value of RadiusSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:displayLabel:RadiusSettings', radiusSettings_displayLabel - Not currently used.

$sel:radiusRetries:RadiusSettings', radiusSettings_radiusRetries - The maximum number of times that communication with the RADIUS server is attempted.

$sel:authenticationProtocol:RadiusSettings', radiusSettings_authenticationProtocol - The protocol specified for your RADIUS endpoints.

$sel:radiusServers:RadiusSettings', radiusSettings_radiusServers - An array of strings that contains the fully qualified domain name (FQDN) or IP addresses of the RADIUS server endpoints, or the FQDN or IP addresses of your RADIUS server load balancer.

$sel:useSameUsername:RadiusSettings', radiusSettings_useSameUsername - Not currently used.

$sel:sharedSecret:RadiusSettings', radiusSettings_sharedSecret - Required for enabling RADIUS on the directory.

$sel:radiusTimeout:RadiusSettings', radiusSettings_radiusTimeout - The amount of time, in seconds, to wait for the RADIUS server to respond.

$sel:radiusPort:RadiusSettings', radiusSettings_radiusPort - The port that your RADIUS server is using for communications. Your self-managed network must allow inbound traffic over this port from the Directory Service servers.

radiusSettings_displayLabel :: Lens' RadiusSettings (Maybe Text)

Not currently used.

radiusSettings_radiusRetries :: Lens' RadiusSettings (Maybe Natural)

The maximum number of times that communication with the RADIUS server is attempted.

radiusSettings_authenticationProtocol :: Lens' RadiusSettings (Maybe RadiusAuthenticationProtocol)

The protocol specified for your RADIUS endpoints.

radiusSettings_radiusServers :: Lens' RadiusSettings (Maybe [Text])

An array of strings that contains the fully qualified domain name (FQDN) or IP addresses of the RADIUS server endpoints, or the FQDN or IP addresses of your RADIUS server load balancer.

radiusSettings_useSameUsername :: Lens' RadiusSettings (Maybe Bool)

Not currently used.

radiusSettings_sharedSecret :: Lens' RadiusSettings (Maybe Text)

Required for enabling RADIUS on the directory.

radiusSettings_radiusTimeout :: Lens' RadiusSettings (Maybe Natural)

The amount of time, in seconds, to wait for the RADIUS server to respond.

radiusSettings_radiusPort :: Lens' RadiusSettings (Maybe Natural)

The port that your RADIUS server is using for communications. Your self-managed network must allow inbound traffic over this port from the Directory Service servers.