Network.AWS.DirectoryService.Types.SchemaExtensionInfo

Description

Documentation

data SchemaExtensionInfo

Information about a schema extension.

See: newSchemaExtensionInfo smart constructor.

Constructors

SchemaExtensionInfo'

Fields

schemaExtensionStatus :: Maybe SchemaExtensionStatus

The current status of the schema extension.

startDateTime :: Maybe POSIX

The date and time that the schema extension started being applied to the directory.

schemaExtensionId :: Maybe Text

The identifier of the schema extension.

directoryId :: Maybe Text

The identifier of the directory to which the schema extension is applied.

description :: Maybe Text

A description of the schema extension.

endDateTime :: Maybe POSIX

The date and time that the schema extension was completed.

schemaExtensionStatusReason :: Maybe Text

The reason for the SchemaExtensionStatus.

newSchemaExtensionInfo :: SchemaExtensionInfo

Create a value of SchemaExtensionInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:schemaExtensionStatus:SchemaExtensionInfo', schemaExtensionInfo_schemaExtensionStatus - The current status of the schema extension.

$sel:startDateTime:SchemaExtensionInfo', schemaExtensionInfo_startDateTime - The date and time that the schema extension started being applied to the directory.

$sel:schemaExtensionId:SchemaExtensionInfo', schemaExtensionInfo_schemaExtensionId - The identifier of the schema extension.

$sel:directoryId:SchemaExtensionInfo', schemaExtensionInfo_directoryId - The identifier of the directory to which the schema extension is applied.

$sel:description:SchemaExtensionInfo', schemaExtensionInfo_description - A description of the schema extension.

$sel:endDateTime:SchemaExtensionInfo', schemaExtensionInfo_endDateTime - The date and time that the schema extension was completed.

$sel:schemaExtensionStatusReason:SchemaExtensionInfo', schemaExtensionInfo_schemaExtensionStatusReason - The reason for the SchemaExtensionStatus.

schemaExtensionInfo_schemaExtensionStatus :: Lens' SchemaExtensionInfo (Maybe SchemaExtensionStatus)

The current status of the schema extension.

schemaExtensionInfo_startDateTime :: Lens' SchemaExtensionInfo (Maybe UTCTime)

The date and time that the schema extension started being applied to the directory.

schemaExtensionInfo_schemaExtensionId :: Lens' SchemaExtensionInfo (Maybe Text)

The identifier of the schema extension.

schemaExtensionInfo_directoryId :: Lens' SchemaExtensionInfo (Maybe Text)

The identifier of the directory to which the schema extension is applied.

schemaExtensionInfo_description :: Lens' SchemaExtensionInfo (Maybe Text)

A description of the schema extension.

schemaExtensionInfo_endDateTime :: Lens' SchemaExtensionInfo (Maybe UTCTime)

The date and time that the schema extension was completed.

schemaExtensionInfo_schemaExtensionStatusReason :: Lens' SchemaExtensionInfo (Maybe Text)

The reason for the SchemaExtensionStatus.