Network.AWS.DirectoryService.Types.Trust

Description

Documentation

data Trust

Describes a trust relationship between an Managed Microsoft AD directory and an external domain.

See: newTrust smart constructor.

Constructors

Trust'

Fields

trustType :: Maybe TrustType

The trust relationship type. Forest is the default.

trustDirection :: Maybe TrustDirection

The trust relationship direction.

stateLastUpdatedDateTime :: Maybe POSIX

The date and time that the TrustState was last updated.

trustId :: Maybe Text

The unique ID of the trust relationship.

createdDateTime :: Maybe POSIX

The date and time that the trust relationship was created.

lastUpdatedDateTime :: Maybe POSIX

The date and time that the trust relationship was last updated.

trustState :: Maybe TrustState

The trust relationship state.

selectiveAuth :: Maybe SelectiveAuth

Current state of selective authentication for the trust.

trustStateReason :: Maybe Text

The reason for the TrustState.

directoryId :: Maybe Text

The Directory ID of the Amazon Web Services directory involved in the trust relationship.

remoteDomainName :: Maybe Text

The Fully Qualified Domain Name (FQDN) of the external domain involved in the trust relationship.

newTrust :: Trust

Create a value of Trust with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:trustType:Trust', trust_trustType - The trust relationship type. Forest is the default.

$sel:trustDirection:Trust', trust_trustDirection - The trust relationship direction.

$sel:stateLastUpdatedDateTime:Trust', trust_stateLastUpdatedDateTime - The date and time that the TrustState was last updated.

$sel:trustId:Trust', trust_trustId - The unique ID of the trust relationship.

$sel:createdDateTime:Trust', trust_createdDateTime - The date and time that the trust relationship was created.

$sel:lastUpdatedDateTime:Trust', trust_lastUpdatedDateTime - The date and time that the trust relationship was last updated.

$sel:trustState:Trust', trust_trustState - The trust relationship state.

$sel:selectiveAuth:Trust', trust_selectiveAuth - Current state of selective authentication for the trust.

$sel:trustStateReason:Trust', trust_trustStateReason - The reason for the TrustState.

$sel:directoryId:Trust', trust_directoryId - The Directory ID of the Amazon Web Services directory involved in the trust relationship.

$sel:remoteDomainName:Trust', trust_remoteDomainName - The Fully Qualified Domain Name (FQDN) of the external domain involved in the trust relationship.

trust_trustType :: Lens' Trust (Maybe TrustType)

The trust relationship type. Forest is the default.

trust_trustDirection :: Lens' Trust (Maybe TrustDirection)

The trust relationship direction.

trust_stateLastUpdatedDateTime :: Lens' Trust (Maybe UTCTime)

The date and time that the TrustState was last updated.

trust_trustId :: Lens' Trust (Maybe Text)

The unique ID of the trust relationship.

trust_createdDateTime :: Lens' Trust (Maybe UTCTime)

The date and time that the trust relationship was created.

trust_lastUpdatedDateTime :: Lens' Trust (Maybe UTCTime)

The date and time that the trust relationship was last updated.

trust_trustState :: Lens' Trust (Maybe TrustState)

The trust relationship state.

trust_selectiveAuth :: Lens' Trust (Maybe SelectiveAuth)

Current state of selective authentication for the trust.

trust_trustStateReason :: Lens' Trust (Maybe Text)

The reason for the TrustState.

trust_directoryId :: Lens' Trust (Maybe Text)

The Directory ID of the Amazon Web Services directory involved in the trust relationship.

trust_remoteDomainName :: Lens' Trust (Maybe Text)

The Fully Qualified Domain Name (FQDN) of the external domain involved in the trust relationship.