Network.AWS.DirectoryService.UnshareDirectory

Description

Stops the directory sharing between the directory owner and consumer accounts.

Creating a Request

data UnshareDirectory

See: newUnshareDirectory smart constructor.

Constructors

UnshareDirectory'

Fields

directoryId :: Text

The identifier of the Managed Microsoft AD directory that you want to stop sharing.

unshareTarget :: UnshareTarget

Identifier for the directory consumer account with whom the directory has to be unshared.

newUnshareDirectory

Arguments

:: Text
$sel:directoryId:UnshareDirectory'

-> UnshareTarget
$sel:unshareTarget:UnshareDirectory'

-> UnshareDirectory

Create a value of UnshareDirectory with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:directoryId:UnshareDirectory', unshareDirectory_directoryId - The identifier of the Managed Microsoft AD directory that you want to stop sharing.

$sel:unshareTarget:UnshareDirectory', unshareDirectory_unshareTarget - Identifier for the directory consumer account with whom the directory has to be unshared.

Request Lenses

unshareDirectory_directoryId :: Lens' UnshareDirectory Text

The identifier of the Managed Microsoft AD directory that you want to stop sharing.

unshareDirectory_unshareTarget :: Lens' UnshareDirectory UnshareTarget

Identifier for the directory consumer account with whom the directory has to be unshared.

Destructuring the Response

data UnshareDirectoryResponse

See: newUnshareDirectoryResponse smart constructor.

Constructors

UnshareDirectoryResponse'

Fields

sharedDirectoryId :: Maybe Text

Identifier of the directory stored in the directory consumer account that is to be unshared from the specified directory (DirectoryId).

httpStatus :: Int

The response's http status code.

newUnshareDirectoryResponse

Arguments

:: Int
$sel:httpStatus:UnshareDirectoryResponse'

-> UnshareDirectoryResponse

Create a value of UnshareDirectoryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sharedDirectoryId:UnshareDirectoryResponse', unshareDirectoryResponse_sharedDirectoryId - Identifier of the directory stored in the directory consumer account that is to be unshared from the specified directory (DirectoryId).

$sel:httpStatus:UnshareDirectoryResponse', unshareDirectoryResponse_httpStatus - The response's http status code.

Response Lenses

unshareDirectoryResponse_sharedDirectoryId :: Lens' UnshareDirectoryResponse (Maybe Text)

Identifier of the directory stored in the directory consumer account that is to be unshared from the specified directory (DirectoryId).

unshareDirectoryResponse_httpStatus :: Lens' UnshareDirectoryResponse Int

The response's http status code.