Network.AWS.DirectoryService.UpdateRadius

Description

Updates the Remote Authentication Dial In User Service (RADIUS) server information for an AD Connector or Microsoft AD directory.

Creating a Request

data UpdateRadius

Contains the inputs for the UpdateRadius operation.

See: newUpdateRadius smart constructor.

Constructors

UpdateRadius'

Fields

directoryId :: Text

The identifier of the directory for which to update the RADIUS server information.

radiusSettings :: RadiusSettings

A RadiusSettings object that contains information about the RADIUS server.

newUpdateRadius

Arguments

:: Text
$sel:directoryId:UpdateRadius'

-> RadiusSettings
$sel:radiusSettings:UpdateRadius'

-> UpdateRadius

Create a value of UpdateRadius with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:directoryId:UpdateRadius', updateRadius_directoryId - The identifier of the directory for which to update the RADIUS server information.

$sel:radiusSettings:UpdateRadius', updateRadius_radiusSettings - A RadiusSettings object that contains information about the RADIUS server.

Request Lenses

updateRadius_directoryId :: Lens' UpdateRadius Text

The identifier of the directory for which to update the RADIUS server information.

updateRadius_radiusSettings :: Lens' UpdateRadius RadiusSettings

A RadiusSettings object that contains information about the RADIUS server.

Destructuring the Response

data UpdateRadiusResponse

Contains the results of the UpdateRadius operation.

See: newUpdateRadiusResponse smart constructor.

Constructors

UpdateRadiusResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateRadiusResponse

Arguments

:: Int
$sel:httpStatus:UpdateRadiusResponse'

-> UpdateRadiusResponse

Create a value of UpdateRadiusResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateRadiusResponse', updateRadiusResponse_httpStatus - The response's http status code.

Response Lenses

updateRadiusResponse_httpStatus :: Lens' UpdateRadiusResponse Int

The response's http status code.