Network.AWS.Discovery.CreateApplication

Description

Creates an application with the given name and description.

Creating a Request

data CreateApplication

See: newCreateApplication smart constructor.

Constructors

CreateApplication'

Fields

description :: Maybe Text

Description of the application to be created.

name :: Text

Name of the application to be created.

newCreateApplication

Arguments

:: Text
$sel:name:CreateApplication'

-> CreateApplication

Create a value of CreateApplication with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:CreateApplication', createApplication_description - Description of the application to be created.

$sel:name:CreateApplication', createApplication_name - Name of the application to be created.

Request Lenses

createApplication_description :: Lens' CreateApplication (Maybe Text)

Description of the application to be created.

createApplication_name :: Lens' CreateApplication Text

Name of the application to be created.

Destructuring the Response

data CreateApplicationResponse

See: newCreateApplicationResponse smart constructor.

Constructors

CreateApplicationResponse'

Fields

configurationId :: Maybe Text

Configuration ID of an application to be created.

httpStatus :: Int

The response's http status code.

newCreateApplicationResponse

Arguments

:: Int
$sel:httpStatus:CreateApplicationResponse'

-> CreateApplicationResponse

Create a value of CreateApplicationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configurationId:CreateApplicationResponse', createApplicationResponse_configurationId - Configuration ID of an application to be created.

$sel:httpStatus:CreateApplicationResponse', createApplicationResponse_httpStatus - The response's http status code.

Response Lenses

createApplicationResponse_configurationId :: Lens' CreateApplicationResponse (Maybe Text)

Configuration ID of an application to be created.

createApplicationResponse_httpStatus :: Lens' CreateApplicationResponse Int

The response's http status code.