Network.AWS.Discovery.DescribeConfigurations

Description

Retrieves attributes for a list of configuration item IDs.

All of the supplied IDs must be for the same asset type from one of the following:

server

application

process

connection

Output fields are specific to the asset type specified. For example, the output for a server configuration item includes a list of attributes about the server, such as host name, operating system, number of network cards, etc.

For a complete list of outputs for each asset type, see Using the DescribeConfigurations Action in the AWS Application Discovery Service User Guide.

Creating a Request

data DescribeConfigurations

See: newDescribeConfigurations smart constructor.

Constructors

DescribeConfigurations'

Fields

configurationIds :: [Text]

One or more configuration IDs.

newDescribeConfigurations :: DescribeConfigurations

Create a value of DescribeConfigurations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configurationIds:DescribeConfigurations', describeConfigurations_configurationIds - One or more configuration IDs.

Request Lenses

describeConfigurations_configurationIds :: Lens' DescribeConfigurations [Text]

One or more configuration IDs.

Destructuring the Response

data DescribeConfigurationsResponse

See: newDescribeConfigurationsResponse smart constructor.

Constructors

DescribeConfigurationsResponse'

Fields

configurations :: Maybe [HashMap Text Text]

A key in the response map. The value is an array of data.

httpStatus :: Int

The response's http status code.

newDescribeConfigurationsResponse

Arguments

:: Int
$sel:httpStatus:DescribeConfigurationsResponse'

-> DescribeConfigurationsResponse

Create a value of DescribeConfigurationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configurations:DescribeConfigurationsResponse', describeConfigurationsResponse_configurations - A key in the response map. The value is an array of data.

$sel:httpStatus:DescribeConfigurationsResponse', describeConfigurationsResponse_httpStatus - The response's http status code.

Response Lenses

describeConfigurationsResponse_configurations :: Lens' DescribeConfigurationsResponse (Maybe [HashMap Text Text])

A key in the response map. The value is an array of data.

describeConfigurationsResponse_httpStatus :: Lens' DescribeConfigurationsResponse Int

The response's http status code.