Network.AWS.Discovery.DescribeContinuousExports

Description

Lists exports as specified by ID. All continuous exports associated with your user account can be listed if you call DescribeContinuousExports as is without passing any parameters.

This operation returns paginated results.

Creating a Request

data DescribeContinuousExports

See: newDescribeContinuousExports smart constructor.

Constructors

DescribeContinuousExports'

Fields

nextToken :: Maybe Text

The token from the previous call to DescribeExportTasks.

maxResults :: Maybe Natural

A number between 1 and 100 specifying the maximum number of continuous export descriptions returned.

exportIds :: Maybe [Text]

The unique IDs assigned to the exports.

newDescribeContinuousExports :: DescribeContinuousExports

Create a value of DescribeContinuousExports with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeContinuousExports', describeContinuousExports_nextToken - The token from the previous call to DescribeExportTasks.

$sel:maxResults:DescribeContinuousExports', describeContinuousExports_maxResults - A number between 1 and 100 specifying the maximum number of continuous export descriptions returned.

$sel:exportIds:DescribeContinuousExports', describeContinuousExports_exportIds - The unique IDs assigned to the exports.

Request Lenses

describeContinuousExports_nextToken :: Lens' DescribeContinuousExports (Maybe Text)

The token from the previous call to DescribeExportTasks.

describeContinuousExports_maxResults :: Lens' DescribeContinuousExports (Maybe Natural)

A number between 1 and 100 specifying the maximum number of continuous export descriptions returned.

describeContinuousExports_exportIds :: Lens' DescribeContinuousExports (Maybe [Text])

The unique IDs assigned to the exports.

Destructuring the Response

data DescribeContinuousExportsResponse

See: newDescribeContinuousExportsResponse smart constructor.

Constructors

DescribeContinuousExportsResponse'

Fields

nextToken :: Maybe Text

The token from the previous call to DescribeExportTasks.

descriptions :: Maybe [ContinuousExportDescription]

A list of continuous export descriptions.

httpStatus :: Int

The response's http status code.

newDescribeContinuousExportsResponse

Arguments

:: Int
$sel:httpStatus:DescribeContinuousExportsResponse'

-> DescribeContinuousExportsResponse

Create a value of DescribeContinuousExportsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeContinuousExports', describeContinuousExportsResponse_nextToken - The token from the previous call to DescribeExportTasks.

$sel:descriptions:DescribeContinuousExportsResponse', describeContinuousExportsResponse_descriptions - A list of continuous export descriptions.

$sel:httpStatus:DescribeContinuousExportsResponse', describeContinuousExportsResponse_httpStatus - The response's http status code.

Response Lenses

describeContinuousExportsResponse_nextToken :: Lens' DescribeContinuousExportsResponse (Maybe Text)

The token from the previous call to DescribeExportTasks.

describeContinuousExportsResponse_descriptions :: Lens' DescribeContinuousExportsResponse (Maybe [ContinuousExportDescription])

A list of continuous export descriptions.

describeContinuousExportsResponse_httpStatus :: Lens' DescribeContinuousExportsResponse Int

The response's http status code.