Network.AWS.Discovery.GetDiscoverySummary

Description

Retrieves a short summary of discovered assets.

This API operation takes no request parameters and is called as is at the command prompt as shown in the example.

Creating a Request

data GetDiscoverySummary

See: newGetDiscoverySummary smart constructor.

Constructors

GetDiscoverySummary'

newGetDiscoverySummary :: GetDiscoverySummary

Create a value of GetDiscoverySummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

Destructuring the Response

data GetDiscoverySummaryResponse

See: newGetDiscoverySummaryResponse smart constructor.

Constructors

GetDiscoverySummaryResponse'

Fields

servers :: Maybe Integer

The number of servers discovered.

agentSummary :: Maybe CustomerAgentInfo

Details about discovered agents, including agent status and health.

connectorSummary :: Maybe CustomerConnectorInfo

Details about discovered connectors, including connector status and health.

applications :: Maybe Integer

The number of applications discovered.

serversMappedToApplications :: Maybe Integer

The number of servers mapped to applications.

serversMappedtoTags :: Maybe Integer

The number of servers mapped to tags.

httpStatus :: Int

The response's http status code.

newGetDiscoverySummaryResponse

Arguments

:: Int
$sel:httpStatus:GetDiscoverySummaryResponse'

-> GetDiscoverySummaryResponse

Create a value of GetDiscoverySummaryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:servers:GetDiscoverySummaryResponse', getDiscoverySummaryResponse_servers - The number of servers discovered.

$sel:agentSummary:GetDiscoverySummaryResponse', getDiscoverySummaryResponse_agentSummary - Details about discovered agents, including agent status and health.

$sel:connectorSummary:GetDiscoverySummaryResponse', getDiscoverySummaryResponse_connectorSummary - Details about discovered connectors, including connector status and health.

$sel:applications:GetDiscoverySummaryResponse', getDiscoverySummaryResponse_applications - The number of applications discovered.

$sel:serversMappedToApplications:GetDiscoverySummaryResponse', getDiscoverySummaryResponse_serversMappedToApplications - The number of servers mapped to applications.

$sel:serversMappedtoTags:GetDiscoverySummaryResponse', getDiscoverySummaryResponse_serversMappedtoTags - The number of servers mapped to tags.

$sel:httpStatus:GetDiscoverySummaryResponse', getDiscoverySummaryResponse_httpStatus - The response's http status code.

Response Lenses

getDiscoverySummaryResponse_servers :: Lens' GetDiscoverySummaryResponse (Maybe Integer)

The number of servers discovered.

getDiscoverySummaryResponse_agentSummary :: Lens' GetDiscoverySummaryResponse (Maybe CustomerAgentInfo)

Details about discovered agents, including agent status and health.

getDiscoverySummaryResponse_connectorSummary :: Lens' GetDiscoverySummaryResponse (Maybe CustomerConnectorInfo)

Details about discovered connectors, including connector status and health.

getDiscoverySummaryResponse_applications :: Lens' GetDiscoverySummaryResponse (Maybe Integer)

The number of applications discovered.

getDiscoverySummaryResponse_serversMappedToApplications :: Lens' GetDiscoverySummaryResponse (Maybe Integer)

The number of servers mapped to applications.

getDiscoverySummaryResponse_serversMappedtoTags :: Lens' GetDiscoverySummaryResponse (Maybe Integer)

The number of servers mapped to tags.

getDiscoverySummaryResponse_httpStatus :: Lens' GetDiscoverySummaryResponse Int

The response's http status code.