Network.AWS.Discovery.Lens

Description

Operations

DescribeExportTasks

describeExportTasks_nextToken :: Lens' DescribeExportTasks (Maybe Text)

The nextToken value returned from a previous paginated DescribeExportTasks request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return.

describeExportTasks_maxResults :: Lens' DescribeExportTasks (Maybe Int)

The maximum number of volume results returned by DescribeExportTasks in paginated output. When this parameter is used, DescribeExportTasks only returns maxResults results in a single page along with a nextToken response element.

describeExportTasks_exportIds :: Lens' DescribeExportTasks (Maybe [Text])

One or more unique identifiers used to query the status of an export request.

describeExportTasks_filters :: Lens' DescribeExportTasks (Maybe [ExportFilter])

One or more filters.

AgentId - ID of the agent whose collected data will be exported

describeExportTasksResponse_nextToken :: Lens' DescribeExportTasksResponse (Maybe Text)

The nextToken value to include in a future DescribeExportTasks request. When the results of a DescribeExportTasks request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

describeExportTasksResponse_exportsInfo :: Lens' DescribeExportTasksResponse (Maybe [ExportInfo])

Contains one or more sets of export request details. When the status of a request is SUCCEEDED, the response includes a URL for an Amazon S3 bucket where you can view the data in a CSV file.

describeExportTasksResponse_httpStatus :: Lens' DescribeExportTasksResponse Int

The response's http status code.

DescribeTags

describeTags_nextToken :: Lens' DescribeTags (Maybe Text)

A token to start the list. Use this token to get the next set of results.

describeTags_maxResults :: Lens' DescribeTags (Maybe Int)

The total number of items to return in a single page of output. The maximum value is 100.

describeTags_filters :: Lens' DescribeTags (Maybe [TagFilter])

You can filter the list using a key-value format. You can separate these items by using logical operators. Allowed filters include tagKey, tagValue, and configurationId.

describeTagsResponse_nextToken :: Lens' DescribeTagsResponse (Maybe Text)

The call returns a token. Use this token to get the next set of results.

describeTagsResponse_tags :: Lens' DescribeTagsResponse (Maybe [ConfigurationTag])

Depending on the input, this is a list of configuration items tagged with a specific tag, or a list of tags for a specific configuration item.

describeTagsResponse_httpStatus :: Lens' DescribeTagsResponse Int

The response's http status code.

ListConfigurations

listConfigurations_nextToken :: Lens' ListConfigurations (Maybe Text)

Token to retrieve the next set of results. For example, if a previous call to ListConfigurations returned 100 items, but you set ListConfigurationsRequest$maxResults to 10, you received a set of 10 results along with a token. Use that token in this query to get the next set of 10.

listConfigurations_maxResults :: Lens' ListConfigurations (Maybe Int)

The total number of items to return. The maximum value is 100.

listConfigurations_orderBy :: Lens' ListConfigurations (Maybe [OrderByElement])

Certain filter criteria return output that can be sorted in ascending or descending order. For a list of output characteristics for each filter, see Using the ListConfigurations Action in the AWS Application Discovery Service User Guide.

listConfigurations_filters :: Lens' ListConfigurations (Maybe [Filter])

You can filter the request using various logical operators and a key-value format. For example:

{"key": "serverType", "value": "webServer"}

For a complete list of filter options and guidance about using them with this action, see Using the ListConfigurations Action in the AWS Application Discovery Service User Guide.

listConfigurations_configurationType :: Lens' ListConfigurations ConfigurationItemType

A valid configuration identified by Application Discovery Service.

listConfigurationsResponse_nextToken :: Lens' ListConfigurationsResponse (Maybe Text)

Token to retrieve the next set of results. For example, if your call to ListConfigurations returned 100 items, but you set ListConfigurationsRequest$maxResults to 10, you received a set of 10 results along with this token. Use this token in the next query to retrieve the next set of 10.

listConfigurationsResponse_configurations :: Lens' ListConfigurationsResponse (Maybe [HashMap Text Text])

Returns configuration details, including the configuration ID, attribute names, and attribute values.

listConfigurationsResponse_httpStatus :: Lens' ListConfigurationsResponse Int

The response's http status code.

StartExportTask

startExportTask_exportDataFormat :: Lens' StartExportTask (Maybe [ExportDataFormat])

The file format for the returned export data. Default value is CSV. Note: The GRAPHML option has been deprecated.

startExportTask_startTime :: Lens' StartExportTask (Maybe UTCTime)

The start timestamp for exported data from the single Application Discovery Agent selected in the filters. If no value is specified, data is exported starting from the first data collected by the agent.

startExportTask_endTime :: Lens' StartExportTask (Maybe UTCTime)

The end timestamp for exported data from the single Application Discovery Agent selected in the filters. If no value is specified, exported data includes the most recent data collected by the agent.

startExportTask_filters :: Lens' StartExportTask (Maybe [ExportFilter])

If a filter is present, it selects the single agentId of the Application Discovery Agent for which data is exported. The agentId can be found in the results of the DescribeAgents API or CLI. If no filter is present, startTime and endTime are ignored and exported data includes both Agentless Discovery Connector data and summary data from Application Discovery agents.

startExportTaskResponse_exportId :: Lens' StartExportTaskResponse (Maybe Text)

A unique identifier used to query the status of an export request.

startExportTaskResponse_httpStatus :: Lens' StartExportTaskResponse Int

The response's http status code.

DescribeImportTasks

describeImportTasks_nextToken :: Lens' DescribeImportTasks (Maybe Text)

The token to request a specific page of results.

describeImportTasks_maxResults :: Lens' DescribeImportTasks (Maybe Natural)

The maximum number of results that you want this request to return, up to 100.

describeImportTasks_filters :: Lens' DescribeImportTasks (Maybe [ImportTaskFilter])

An array of name-value pairs that you provide to filter the results for the DescribeImportTask request to a specific subset of results. Currently, wildcard values aren't supported for filters.

describeImportTasksResponse_nextToken :: Lens' DescribeImportTasksResponse (Maybe Text)

The token to request the next page of results.

describeImportTasksResponse_tasks :: Lens' DescribeImportTasksResponse (Maybe [ImportTask])

A returned array of import tasks that match any applied filters, up to the specified number of maximum results.

describeImportTasksResponse_httpStatus :: Lens' DescribeImportTasksResponse Int

The response's http status code.

CreateApplication

createApplication_description :: Lens' CreateApplication (Maybe Text)

Description of the application to be created.

createApplication_name :: Lens' CreateApplication Text

Name of the application to be created.

createApplicationResponse_configurationId :: Lens' CreateApplicationResponse (Maybe Text)

Configuration ID of an application to be created.

createApplicationResponse_httpStatus :: Lens' CreateApplicationResponse Int

The response's http status code.

DisassociateConfigurationItemsFromApplication

disassociateConfigurationItemsFromApplication_applicationConfigurationId :: Lens' DisassociateConfigurationItemsFromApplication Text

Configuration ID of an application from which each item is disassociated.

disassociateConfigurationItemsFromApplication_configurationIds :: Lens' DisassociateConfigurationItemsFromApplication [Text]

Configuration ID of each item to be disassociated from an application.

disassociateConfigurationItemsFromApplicationResponse_httpStatus :: Lens' DisassociateConfigurationItemsFromApplicationResponse Int

The response's http status code.

DeleteTags

deleteTags_tags :: Lens' DeleteTags (Maybe [Tag])

Tags that you want to delete from one or more configuration items. Specify the tags that you want to delete in a key-value format. For example:

{"key": "serverType", "value": "webServer"}

deleteTags_configurationIds :: Lens' DeleteTags [Text]

A list of configuration items with tags that you want to delete.

deleteTagsResponse_httpStatus :: Lens' DeleteTagsResponse Int

The response's http status code.

GetDiscoverySummary

getDiscoverySummaryResponse_servers :: Lens' GetDiscoverySummaryResponse (Maybe Integer)

The number of servers discovered.

getDiscoverySummaryResponse_agentSummary :: Lens' GetDiscoverySummaryResponse (Maybe CustomerAgentInfo)

Details about discovered agents, including agent status and health.

getDiscoverySummaryResponse_connectorSummary :: Lens' GetDiscoverySummaryResponse (Maybe CustomerConnectorInfo)

Details about discovered connectors, including connector status and health.

getDiscoverySummaryResponse_applications :: Lens' GetDiscoverySummaryResponse (Maybe Integer)

The number of applications discovered.

getDiscoverySummaryResponse_serversMappedToApplications :: Lens' GetDiscoverySummaryResponse (Maybe Integer)

The number of servers mapped to applications.

getDiscoverySummaryResponse_serversMappedtoTags :: Lens' GetDiscoverySummaryResponse (Maybe Integer)

The number of servers mapped to tags.

getDiscoverySummaryResponse_httpStatus :: Lens' GetDiscoverySummaryResponse Int

The response's http status code.

DeleteApplications

deleteApplications_configurationIds :: Lens' DeleteApplications [Text]

Configuration ID of an application to be deleted.

deleteApplicationsResponse_httpStatus :: Lens' DeleteApplicationsResponse Int

The response's http status code.

StartDataCollectionByAgentIds

startDataCollectionByAgentIds_agentIds :: Lens' StartDataCollectionByAgentIds [Text]

The IDs of the agents or connectors from which to start collecting data. If you send a request to an agent/connector ID that you do not have permission to contact, according to your AWS account, the service does not throw an exception. Instead, it returns the error in the Description field. If you send a request to multiple agents/connectors and you do not have permission to contact some of those agents/connectors, the system does not throw an exception. Instead, the system shows Failed in the Description field.

startDataCollectionByAgentIdsResponse_agentsConfigurationStatus :: Lens' StartDataCollectionByAgentIdsResponse (Maybe [AgentConfigurationStatus])

Information about agents or the connector that were instructed to start collecting data. Information includes the agent/connector ID, a description of the operation performed, and whether the agent/connector configuration was updated.

startDataCollectionByAgentIdsResponse_httpStatus :: Lens' StartDataCollectionByAgentIdsResponse Int

The response's http status code.

StopDataCollectionByAgentIds

stopDataCollectionByAgentIds_agentIds :: Lens' StopDataCollectionByAgentIds [Text]

The IDs of the agents or connectors from which to stop collecting data.

stopDataCollectionByAgentIdsResponse_agentsConfigurationStatus :: Lens' StopDataCollectionByAgentIdsResponse (Maybe [AgentConfigurationStatus])

Information about the agents or connector that were instructed to stop collecting data. Information includes the agent/connector ID, a description of the operation performed, and whether the agent/connector configuration was updated.

stopDataCollectionByAgentIdsResponse_httpStatus :: Lens' StopDataCollectionByAgentIdsResponse Int

The response's http status code.

DescribeContinuousExports

describeContinuousExports_nextToken :: Lens' DescribeContinuousExports (Maybe Text)

The token from the previous call to DescribeExportTasks.

describeContinuousExports_maxResults :: Lens' DescribeContinuousExports (Maybe Natural)

A number between 1 and 100 specifying the maximum number of continuous export descriptions returned.

describeContinuousExports_exportIds :: Lens' DescribeContinuousExports (Maybe [Text])

The unique IDs assigned to the exports.

describeContinuousExportsResponse_nextToken :: Lens' DescribeContinuousExportsResponse (Maybe Text)

The token from the previous call to DescribeExportTasks.

describeContinuousExportsResponse_descriptions :: Lens' DescribeContinuousExportsResponse (Maybe [ContinuousExportDescription])

A list of continuous export descriptions.

describeContinuousExportsResponse_httpStatus :: Lens' DescribeContinuousExportsResponse Int

The response's http status code.

DescribeAgents

describeAgents_agentIds :: Lens' DescribeAgents (Maybe [Text])

The agent or the Connector IDs for which you want information. If you specify no IDs, the system returns information about all agents/Connectors associated with your AWS user account.

describeAgents_nextToken :: Lens' DescribeAgents (Maybe Text)

Token to retrieve the next set of results. For example, if you previously specified 100 IDs for DescribeAgentsRequest$agentIds but set DescribeAgentsRequest$maxResults to 10, you received a set of 10 results along with a token. Use that token in this query to get the next set of 10.

describeAgents_maxResults :: Lens' DescribeAgents (Maybe Int)

The total number of agents/Connectors to return in a single page of output. The maximum value is 100.

describeAgents_filters :: Lens' DescribeAgents (Maybe [Filter])

You can filter the request using various logical operators and a key-value format. For example:

{"key": "collectionStatus", "value": "STARTED"}

describeAgentsResponse_nextToken :: Lens' DescribeAgentsResponse (Maybe Text)

Token to retrieve the next set of results. For example, if you specified 100 IDs for DescribeAgentsRequest$agentIds but set DescribeAgentsRequest$maxResults to 10, you received a set of 10 results along with this token. Use this token in the next query to retrieve the next set of 10.

describeAgentsResponse_agentsInfo :: Lens' DescribeAgentsResponse (Maybe [AgentInfo])

Lists agents or the Connector by ID or lists all agents/Connectors associated with your user account if you did not specify an agent/Connector ID. The output includes agent/Connector IDs, IP addresses, media access control (MAC) addresses, agent/Connector health, host name where the agent/Connector resides, and the version number of each agent/Connector.

describeAgentsResponse_httpStatus :: Lens' DescribeAgentsResponse Int

The response's http status code.

StartContinuousExport

startContinuousExportResponse_s3Bucket :: Lens' StartContinuousExportResponse (Maybe Text)

The name of the s3 bucket where the export data parquet files are stored.

startContinuousExportResponse_startTime :: Lens' StartContinuousExportResponse (Maybe UTCTime)

The timestamp representing when the continuous export was started.

startContinuousExportResponse_dataSource :: Lens' StartContinuousExportResponse (Maybe DataSource)

The type of data collector used to gather this data (currently only offered for AGENT).

startContinuousExportResponse_schemaStorageConfig :: Lens' StartContinuousExportResponse (Maybe (HashMap Text Text))

A dictionary which describes how the data is stored.

databaseName - the name of the Glue database used to store the schema.

startContinuousExportResponse_exportId :: Lens' StartContinuousExportResponse (Maybe Text)

The unique ID assigned to this export.

startContinuousExportResponse_httpStatus :: Lens' StartContinuousExportResponse Int

The response's http status code.

StopContinuousExport

stopContinuousExport_exportId :: Lens' StopContinuousExport Text

The unique ID assigned to this export.

stopContinuousExportResponse_stopTime :: Lens' StopContinuousExportResponse (Maybe UTCTime)

Timestamp that represents when this continuous export was stopped.

stopContinuousExportResponse_startTime :: Lens' StopContinuousExportResponse (Maybe UTCTime)

Timestamp that represents when this continuous export started collecting data.

stopContinuousExportResponse_httpStatus :: Lens' StopContinuousExportResponse Int

The response's http status code.

ListServerNeighbors

listServerNeighbors_nextToken :: Lens' ListServerNeighbors (Maybe Text)

Token to retrieve the next set of results. For example, if you previously specified 100 IDs for ListServerNeighborsRequest$neighborConfigurationIds but set ListServerNeighborsRequest$maxResults to 10, you received a set of 10 results along with a token. Use that token in this query to get the next set of 10.

listServerNeighbors_maxResults :: Lens' ListServerNeighbors (Maybe Int)

Maximum number of results to return in a single page of output.

listServerNeighbors_portInformationNeeded :: Lens' ListServerNeighbors (Maybe Bool)

Flag to indicate if port and protocol information is needed as part of the response.

listServerNeighbors_neighborConfigurationIds :: Lens' ListServerNeighbors (Maybe [Text])

List of configuration IDs to test for one-hop-away.

listServerNeighbors_configurationId :: Lens' ListServerNeighbors Text

Configuration ID of the server for which neighbors are being listed.

listServerNeighborsResponse_nextToken :: Lens' ListServerNeighborsResponse (Maybe Text)

Token to retrieve the next set of results. For example, if you specified 100 IDs for ListServerNeighborsRequest$neighborConfigurationIds but set ListServerNeighborsRequest$maxResults to 10, you received a set of 10 results along with this token. Use this token in the next query to retrieve the next set of 10.

listServerNeighborsResponse_knownDependencyCount :: Lens' ListServerNeighborsResponse (Maybe Integer)

Count of distinct servers that are one hop away from the given server.

listServerNeighborsResponse_httpStatus :: Lens' ListServerNeighborsResponse Int

The response's http status code.

listServerNeighborsResponse_neighbors :: Lens' ListServerNeighborsResponse [NeighborConnectionDetail]

List of distinct servers that are one hop away from the given server.

DescribeConfigurations

describeConfigurations_configurationIds :: Lens' DescribeConfigurations [Text]

One or more configuration IDs.

describeConfigurationsResponse_configurations :: Lens' DescribeConfigurationsResponse (Maybe [HashMap Text Text])

A key in the response map. The value is an array of data.

describeConfigurationsResponse_httpStatus :: Lens' DescribeConfigurationsResponse Int

The response's http status code.

AssociateConfigurationItemsToApplication

associateConfigurationItemsToApplication_applicationConfigurationId :: Lens' AssociateConfigurationItemsToApplication Text

The configuration ID of an application with which items are to be associated.

associateConfigurationItemsToApplication_configurationIds :: Lens' AssociateConfigurationItemsToApplication [Text]

The ID of each configuration item to be associated with an application.

associateConfigurationItemsToApplicationResponse_httpStatus :: Lens' AssociateConfigurationItemsToApplicationResponse Int

The response's http status code.

UpdateApplication

updateApplication_name :: Lens' UpdateApplication (Maybe Text)

New name of the application to be updated.

updateApplication_description :: Lens' UpdateApplication (Maybe Text)

New description of the application to be updated.

updateApplication_configurationId :: Lens' UpdateApplication Text

Configuration ID of the application to be updated.

updateApplicationResponse_httpStatus :: Lens' UpdateApplicationResponse Int

The response's http status code.

StartImportTask

startImportTask_clientRequestToken :: Lens' StartImportTask (Maybe Text)

Optional. A unique token that you can provide to prevent the same import request from occurring more than once. If you don't provide a token, a token is automatically generated.

Sending more than one StartImportTask request with the same client request token will return information about the original import task with that client request token.

startImportTask_name :: Lens' StartImportTask Text

A descriptive name for this request. You can use this name to filter future requests related to this import task, such as identifying applications and servers that were included in this import task. We recommend that you use a meaningful name for each import task.

startImportTask_importUrl :: Lens' StartImportTask Text

The URL for your import file that you've uploaded to Amazon S3.

If you're using the AWS CLI, this URL is structured as follows: s3://BucketName/ImportFileName.CSV

startImportTaskResponse_task :: Lens' StartImportTaskResponse (Maybe ImportTask)

An array of information related to the import task request including status information, times, IDs, the Amazon S3 Object URL for the import file, and more.

startImportTaskResponse_httpStatus :: Lens' StartImportTaskResponse Int

The response's http status code.

BatchDeleteImportData

batchDeleteImportData_importTaskIds :: Lens' BatchDeleteImportData (NonEmpty Text)

The IDs for the import tasks that you want to delete.

batchDeleteImportDataResponse_errors :: Lens' BatchDeleteImportDataResponse (Maybe [BatchDeleteImportDataError])

Error messages returned for each import task that you deleted as a response for this command.

batchDeleteImportDataResponse_httpStatus :: Lens' BatchDeleteImportDataResponse Int

The response's http status code.

CreateTags

createTags_configurationIds :: Lens' CreateTags [Text]

A list of configuration items that you want to tag.

createTags_tags :: Lens' CreateTags [Tag]

Tags that you want to associate with one or more configuration items. Specify the tags that you want to create in a key-value format. For example:

{"key": "serverType", "value": "webServer"}

createTagsResponse_httpStatus :: Lens' CreateTagsResponse Int

The response's http status code.

Types

AgentConfigurationStatus

agentConfigurationStatus_agentId :: Lens' AgentConfigurationStatus (Maybe Text)

The agent/connector ID.

agentConfigurationStatus_operationSucceeded :: Lens' AgentConfigurationStatus (Maybe Bool)

Information about the status of the StartDataCollection and StopDataCollection operations. The system has recorded the data collection operation. The agent/connector receives this command the next time it polls for a new command.

agentConfigurationStatus_description :: Lens' AgentConfigurationStatus (Maybe Text)

A description of the operation performed.

AgentInfo

agentInfo_hostName :: Lens' AgentInfo (Maybe Text)

The name of the host where the agent or connector resides. The host can be a server or virtual machine.

agentInfo_agentId :: Lens' AgentInfo (Maybe Text)

The agent or connector ID.

agentInfo_agentType :: Lens' AgentInfo (Maybe Text)

Type of agent.

agentInfo_connectorId :: Lens' AgentInfo (Maybe Text)

The ID of the connector.

agentInfo_lastHealthPingTime :: Lens' AgentInfo (Maybe Text)

Time since agent or connector health was reported.

agentInfo_agentNetworkInfoList :: Lens' AgentInfo (Maybe [AgentNetworkInfo])

Network details about the host where the agent or connector resides.

agentInfo_registeredTime :: Lens' AgentInfo (Maybe Text)

Agent's first registration timestamp in UTC.

agentInfo_version :: Lens' AgentInfo (Maybe Text)

The agent or connector version.

agentInfo_health :: Lens' AgentInfo (Maybe AgentStatus)

The health of the agent or connector.

agentInfo_collectionStatus :: Lens' AgentInfo (Maybe Text)

Status of the collection process for an agent or connector.

AgentNetworkInfo

agentNetworkInfo_macAddress :: Lens' AgentNetworkInfo (Maybe Text)

The MAC address for the host where the agent/connector resides.

agentNetworkInfo_ipAddress :: Lens' AgentNetworkInfo (Maybe Text)

The IP address for the host where the agent/connector resides.

BatchDeleteImportDataError

batchDeleteImportDataError_errorDescription :: Lens' BatchDeleteImportDataError (Maybe Text)

The description of the error that occurred for a specific import task.

batchDeleteImportDataError_importTaskId :: Lens' BatchDeleteImportDataError (Maybe Text)

The unique import ID associated with the error that occurred.

batchDeleteImportDataError_errorCode :: Lens' BatchDeleteImportDataError (Maybe BatchDeleteImportDataErrorCode)

The type of error that occurred for a specific import task.

ConfigurationTag

configurationTag_key :: Lens' ConfigurationTag (Maybe Text)

A type of tag on which to filter. For example, serverType.

configurationTag_configurationId :: Lens' ConfigurationTag (Maybe Text)

The configuration ID for the item to tag. You can specify a list of keys and values.

configurationTag_value :: Lens' ConfigurationTag (Maybe Text)

A value on which to filter. For example key = serverType and /value = web server/.

configurationTag_configurationType :: Lens' ConfigurationTag (Maybe ConfigurationItemType)

A type of IT asset to tag.

configurationTag_timeOfCreation :: Lens' ConfigurationTag (Maybe UTCTime)

The time the configuration tag was created in Coordinated Universal Time (UTC).

ContinuousExportDescription

continuousExportDescription_status :: Lens' ContinuousExportDescription (Maybe ContinuousExportStatus)

Describes the status of the export. Can be one of the following values:

START_IN_PROGRESS - setting up resources to start continuous export.

START_FAILED - an error occurred setting up continuous export. To recover, call start-continuous-export again.

ACTIVE - data is being exported to the customer bucket.

ERROR - an error occurred during export. To fix the issue, call stop-continuous-export and start-continuous-export.

STOP_IN_PROGRESS - stopping the export.

STOP_FAILED - an error occurred stopping the export. To recover, call stop-continuous-export again.

INACTIVE - the continuous export has been stopped. Data is no longer being exported to the customer bucket.

continuousExportDescription_s3Bucket :: Lens' ContinuousExportDescription (Maybe Text)

The name of the s3 bucket where the export data parquet files are stored.

continuousExportDescription_stopTime :: Lens' ContinuousExportDescription (Maybe UTCTime)

The timestamp that represents when this continuous export was stopped.

continuousExportDescription_startTime :: Lens' ContinuousExportDescription (Maybe UTCTime)

The timestamp representing when the continuous export was started.

continuousExportDescription_dataSource :: Lens' ContinuousExportDescription (Maybe DataSource)

The type of data collector used to gather this data (currently only offered for AGENT).

continuousExportDescription_statusDetail :: Lens' ContinuousExportDescription (Maybe Text)

Contains information about any errors that have occurred. This data type can have the following values:

ACCESS_DENIED - You don’t have permission to start Data Exploration in Amazon Athena. Contact your AWS administrator for help. For more information, see Setting Up AWS Application Discovery Service in the Application Discovery Service User Guide.

DELIVERY_STREAM_LIMIT_FAILURE - You reached the limit for Amazon Kinesis Data Firehose delivery streams. Reduce the number of streams or request a limit increase and try again. For more information, see Kinesis Data Streams Limits in the Amazon Kinesis Data Streams Developer Guide.

FIREHOSE_ROLE_MISSING - The Data Exploration feature is in an error state because your IAM User is missing the AWSApplicationDiscoveryServiceFirehose role. Turn on Data Exploration in Amazon Athena and try again. For more information, see Step 3: Provide Application Discovery Service Access to Non-Administrator Users by Attaching Policies in the Application Discovery Service User Guide.

FIREHOSE_STREAM_DOES_NOT_EXIST - The Data Exploration feature is in an error state because your IAM User is missing one or more of the Kinesis data delivery streams.

INTERNAL_FAILURE - The Data Exploration feature is in an error state because of an internal failure. Try again later. If this problem persists, contact AWS Support.

S3_BUCKET_LIMIT_FAILURE - You reached the limit for Amazon S3 buckets. Reduce the number of Amazon S3 buckets or request a limit increase and try again. For more information, see Bucket Restrictions and Limitations in the Amazon Simple Storage Service Developer Guide.

S3_NOT_SIGNED_UP - Your account is not signed up for the Amazon S3 service. You must sign up before you can use Amazon S3. You can sign up at the following URL: https://aws.amazon.com/s3.

continuousExportDescription_schemaStorageConfig :: Lens' ContinuousExportDescription (Maybe (HashMap Text Text))

An object which describes how the data is stored.

databaseName - the name of the Glue database used to store the schema.

continuousExportDescription_exportId :: Lens' ContinuousExportDescription (Maybe Text)

The unique ID assigned to this export.

CustomerAgentInfo

customerAgentInfo_activeAgents :: Lens' CustomerAgentInfo Int

Number of active discovery agents.

customerAgentInfo_healthyAgents :: Lens' CustomerAgentInfo Int

Number of healthy discovery agents

customerAgentInfo_blackListedAgents :: Lens' CustomerAgentInfo Int

Number of blacklisted discovery agents.

customerAgentInfo_shutdownAgents :: Lens' CustomerAgentInfo Int

Number of discovery agents with status SHUTDOWN.

customerAgentInfo_unhealthyAgents :: Lens' CustomerAgentInfo Int

Number of unhealthy discovery agents.

customerAgentInfo_totalAgents :: Lens' CustomerAgentInfo Int

Total number of discovery agents.

customerAgentInfo_unknownAgents :: Lens' CustomerAgentInfo Int

Number of unknown discovery agents.

CustomerConnectorInfo

customerConnectorInfo_activeConnectors :: Lens' CustomerConnectorInfo Int

Number of active discovery connectors.

customerConnectorInfo_healthyConnectors :: Lens' CustomerConnectorInfo Int

Number of healthy discovery connectors.

customerConnectorInfo_blackListedConnectors :: Lens' CustomerConnectorInfo Int

Number of blacklisted discovery connectors.

customerConnectorInfo_shutdownConnectors :: Lens' CustomerConnectorInfo Int

Number of discovery connectors with status SHUTDOWN,

customerConnectorInfo_unhealthyConnectors :: Lens' CustomerConnectorInfo Int

Number of unhealthy discovery connectors.

customerConnectorInfo_totalConnectors :: Lens' CustomerConnectorInfo Int

Total number of discovery connectors.

customerConnectorInfo_unknownConnectors :: Lens' CustomerConnectorInfo Int

Number of unknown discovery connectors.

ExportFilter

exportFilter_name :: Lens' ExportFilter Text

A single ExportFilter name. Supported filters: agentId.

exportFilter_values :: Lens' ExportFilter [Text]

A single agentId for a Discovery Agent. An agentId can be found using the DescribeAgents action. Typically an ADS agentId is in the form o-0123456789abcdef0.

exportFilter_condition :: Lens' ExportFilter Text

Supported condition: EQUALS

ExportInfo

exportInfo_isTruncated :: Lens' ExportInfo (Maybe Bool)

If true, the export of agent information exceeded the size limit for a single export and the exported data is incomplete for the requested time range. To address this, select a smaller time range for the export by using startDate and endDate.

exportInfo_configurationsDownloadUrl :: Lens' ExportInfo (Maybe Text)

A URL for an Amazon S3 bucket where you can review the exported data. The URL is displayed only if the export succeeded.

exportInfo_requestedStartTime :: Lens' ExportInfo (Maybe UTCTime)

The value of startTime parameter in the StartExportTask request. If no startTime was requested, this result does not appear in ExportInfo.

exportInfo_requestedEndTime :: Lens' ExportInfo (Maybe UTCTime)

The endTime used in the StartExportTask request. If no endTime was requested, this result does not appear in ExportInfo.

exportInfo_exportId :: Lens' ExportInfo Text

A unique identifier used to query an export.

exportInfo_exportStatus :: Lens' ExportInfo ExportStatus

The status of the data export job.

exportInfo_statusMessage :: Lens' ExportInfo Text

A status message provided for API callers.

exportInfo_exportRequestTime :: Lens' ExportInfo UTCTime

The time that the data export was initiated.

Filter

filter_name :: Lens' Filter Text

The name of the filter.

filter_values :: Lens' Filter [Text]

A string value on which to filter. For example, if you choose the destinationServer.osVersion filter name, you could specify Ubuntu for the value.

filter_condition :: Lens' Filter Text

A conditional operator. The following operators are valid: EQUALS, NOT_EQUALS, CONTAINS, NOT_CONTAINS. If you specify multiple filters, the system utilizes all filters as though concatenated by AND. If you specify multiple values for a particular filter, the system differentiates the values using OR. Calling either DescribeConfigurations or ListConfigurations returns attributes of matching configuration items.

ImportTask

importTask_applicationImportSuccess :: Lens' ImportTask (Maybe Int)

The total number of application records in the import file that were successfully imported.

importTask_status :: Lens' ImportTask (Maybe ImportStatus)

The status of the import task. An import can have the status of IMPORT_COMPLETE and still have some records fail to import from the overall request. More information can be found in the downloadable archive defined in the errorsAndFailedEntriesZip field, or in the Migration Hub management console.

importTask_importRequestTime :: Lens' ImportTask (Maybe UTCTime)

The time that the import task request was made, presented in the Unix time stamp format.

importTask_serverImportFailure :: Lens' ImportTask (Maybe Int)

The total number of server records in the import file that failed to be imported.

importTask_importTaskId :: Lens' ImportTask (Maybe Text)

The unique ID for a specific import task. These IDs aren't globally unique, but they are unique within an AWS account.

importTask_errorsAndFailedEntriesZip :: Lens' ImportTask (Maybe Text)

A link to a compressed archive folder (in the ZIP format) that contains an error log and a file of failed records. You can use these two files to quickly identify records that failed, why they failed, and correct those records. Afterward, you can upload the corrected file to your Amazon S3 bucket and create another import task request.

This field also includes authorization information so you can confirm the authenticity of the compressed archive before you download it.

If some records failed to be imported we recommend that you correct the records in the failed entries file and then imports that failed entries file. This prevents you from having to correct and update the larger original file and attempt importing it again.

importTask_applicationImportFailure :: Lens' ImportTask (Maybe Int)

The total number of application records in the import file that failed to be imported.

importTask_name :: Lens' ImportTask (Maybe Text)

A descriptive name for an import task. You can use this name to filter future requests related to this import task, such as identifying applications and servers that were included in this import task. We recommend that you use a meaningful name for each import task.

importTask_importCompletionTime :: Lens' ImportTask (Maybe UTCTime)

The time that the import task request finished, presented in the Unix time stamp format.

importTask_importUrl :: Lens' ImportTask (Maybe Text)

The URL for your import file that you've uploaded to Amazon S3.

importTask_serverImportSuccess :: Lens' ImportTask (Maybe Int)

The total number of server records in the import file that were successfully imported.

importTask_clientRequestToken :: Lens' ImportTask (Maybe Text)

A unique token used to prevent the same import request from occurring more than once. If you didn't provide a token, a token was automatically generated when the import task request was sent.

importTask_importDeletedTime :: Lens' ImportTask (Maybe UTCTime)

The time that the import task request was deleted, presented in the Unix time stamp format.

ImportTaskFilter

importTaskFilter_values :: Lens' ImportTaskFilter (Maybe (NonEmpty Text))

An array of strings that you can provide to match against a specific name, status, or import task ID to filter the results for your import task queries.

importTaskFilter_name :: Lens' ImportTaskFilter (Maybe ImportTaskFilterName)

The name, status, or import task ID for a specific import task.

NeighborConnectionDetail

neighborConnectionDetail_transportProtocol :: Lens' NeighborConnectionDetail (Maybe Text)

The network protocol used for the connection.

neighborConnectionDetail_destinationPort :: Lens' NeighborConnectionDetail (Maybe Int)

The destination network port for the connection.

neighborConnectionDetail_sourceServerId :: Lens' NeighborConnectionDetail Text

The ID of the server that opened the network connection.

neighborConnectionDetail_destinationServerId :: Lens' NeighborConnectionDetail Text

The ID of the server that accepted the network connection.

neighborConnectionDetail_connectionsCount :: Lens' NeighborConnectionDetail Integer

The number of open network connections with the neighboring server.

OrderByElement

orderByElement_sortOrder :: Lens' OrderByElement (Maybe OrderString)

Ordering direction.

orderByElement_fieldName :: Lens' OrderByElement Text

The field on which to order.

Tag

tag_key :: Lens' Tag Text

The type of tag on which to filter.

tag_value :: Lens' Tag Text

A value for a tag key on which to filter.

TagFilter

tagFilter_name :: Lens' TagFilter Text

A name of the tag filter.

tagFilter_values :: Lens' TagFilter [Text]

Values for the tag filter.