Network.AWS.Discovery.ListServerNeighbors

Description

Retrieves a list of servers that are one network hop away from a specified server.

Creating a Request

data ListServerNeighbors

See: newListServerNeighbors smart constructor.

Constructors

ListServerNeighbors'

Fields

nextToken :: Maybe Text

Token to retrieve the next set of results. For example, if you previously specified 100 IDs for ListServerNeighborsRequest$neighborConfigurationIds but set ListServerNeighborsRequest$maxResults to 10, you received a set of 10 results along with a token. Use that token in this query to get the next set of 10.

maxResults :: Maybe Int

Maximum number of results to return in a single page of output.

portInformationNeeded :: Maybe Bool

Flag to indicate if port and protocol information is needed as part of the response.

neighborConfigurationIds :: Maybe [Text]

List of configuration IDs to test for one-hop-away.

configurationId :: Text

Configuration ID of the server for which neighbors are being listed.

newListServerNeighbors

Arguments

:: Text
$sel:configurationId:ListServerNeighbors'

-> ListServerNeighbors

Create a value of ListServerNeighbors with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListServerNeighbors', listServerNeighbors_nextToken - Token to retrieve the next set of results. For example, if you previously specified 100 IDs for ListServerNeighborsRequest$neighborConfigurationIds but set ListServerNeighborsRequest$maxResults to 10, you received a set of 10 results along with a token. Use that token in this query to get the next set of 10.

$sel:maxResults:ListServerNeighbors', listServerNeighbors_maxResults - Maximum number of results to return in a single page of output.

$sel:portInformationNeeded:ListServerNeighbors', listServerNeighbors_portInformationNeeded - Flag to indicate if port and protocol information is needed as part of the response.

$sel:neighborConfigurationIds:ListServerNeighbors', listServerNeighbors_neighborConfigurationIds - List of configuration IDs to test for one-hop-away.

$sel:configurationId:ListServerNeighbors', listServerNeighbors_configurationId - Configuration ID of the server for which neighbors are being listed.

Request Lenses

listServerNeighbors_nextToken :: Lens' ListServerNeighbors (Maybe Text)

Token to retrieve the next set of results. For example, if you previously specified 100 IDs for ListServerNeighborsRequest$neighborConfigurationIds but set ListServerNeighborsRequest$maxResults to 10, you received a set of 10 results along with a token. Use that token in this query to get the next set of 10.

listServerNeighbors_maxResults :: Lens' ListServerNeighbors (Maybe Int)

Maximum number of results to return in a single page of output.

listServerNeighbors_portInformationNeeded :: Lens' ListServerNeighbors (Maybe Bool)

Flag to indicate if port and protocol information is needed as part of the response.

listServerNeighbors_neighborConfigurationIds :: Lens' ListServerNeighbors (Maybe [Text])

List of configuration IDs to test for one-hop-away.

listServerNeighbors_configurationId :: Lens' ListServerNeighbors Text

Configuration ID of the server for which neighbors are being listed.

Destructuring the Response

data ListServerNeighborsResponse

See: newListServerNeighborsResponse smart constructor.

Constructors

ListServerNeighborsResponse'

Fields

nextToken :: Maybe Text

Token to retrieve the next set of results. For example, if you specified 100 IDs for ListServerNeighborsRequest$neighborConfigurationIds but set ListServerNeighborsRequest$maxResults to 10, you received a set of 10 results along with this token. Use this token in the next query to retrieve the next set of 10.

knownDependencyCount :: Maybe Integer

Count of distinct servers that are one hop away from the given server.

httpStatus :: Int

The response's http status code.

neighbors :: [NeighborConnectionDetail]

List of distinct servers that are one hop away from the given server.

newListServerNeighborsResponse

Arguments

:: Int
$sel:httpStatus:ListServerNeighborsResponse'

-> ListServerNeighborsResponse

Create a value of ListServerNeighborsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListServerNeighbors', listServerNeighborsResponse_nextToken - Token to retrieve the next set of results. For example, if you specified 100 IDs for ListServerNeighborsRequest$neighborConfigurationIds but set ListServerNeighborsRequest$maxResults to 10, you received a set of 10 results along with this token. Use this token in the next query to retrieve the next set of 10.

$sel:knownDependencyCount:ListServerNeighborsResponse', listServerNeighborsResponse_knownDependencyCount - Count of distinct servers that are one hop away from the given server.

$sel:httpStatus:ListServerNeighborsResponse', listServerNeighborsResponse_httpStatus - The response's http status code.

$sel:neighbors:ListServerNeighborsResponse', listServerNeighborsResponse_neighbors - List of distinct servers that are one hop away from the given server.

Response Lenses

listServerNeighborsResponse_nextToken :: Lens' ListServerNeighborsResponse (Maybe Text)

Token to retrieve the next set of results. For example, if you specified 100 IDs for ListServerNeighborsRequest$neighborConfigurationIds but set ListServerNeighborsRequest$maxResults to 10, you received a set of 10 results along with this token. Use this token in the next query to retrieve the next set of 10.

listServerNeighborsResponse_knownDependencyCount :: Lens' ListServerNeighborsResponse (Maybe Integer)

Count of distinct servers that are one hop away from the given server.

listServerNeighborsResponse_httpStatus :: Lens' ListServerNeighborsResponse Int

The response's http status code.

listServerNeighborsResponse_neighbors :: Lens' ListServerNeighborsResponse [NeighborConnectionDetail]

List of distinct servers that are one hop away from the given server.