Network.AWS.Discovery.StartContinuousExport

Description

Start the continuous flow of agent's discovered data into Amazon Athena.

Creating a Request

data StartContinuousExport

See: newStartContinuousExport smart constructor.

Constructors

StartContinuousExport'

newStartContinuousExport :: StartContinuousExport

Create a value of StartContinuousExport with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

Destructuring the Response

data StartContinuousExportResponse

See: newStartContinuousExportResponse smart constructor.

Constructors

StartContinuousExportResponse'

Fields

s3Bucket :: Maybe Text

The name of the s3 bucket where the export data parquet files are stored.

startTime :: Maybe POSIX

The timestamp representing when the continuous export was started.

dataSource :: Maybe DataSource

The type of data collector used to gather this data (currently only offered for AGENT).

schemaStorageConfig :: Maybe (HashMap Text Text)

A dictionary which describes how the data is stored.

databaseName - the name of the Glue database used to store the schema.

exportId :: Maybe Text

The unique ID assigned to this export.

httpStatus :: Int

The response's http status code.

newStartContinuousExportResponse

Arguments

:: Int
$sel:httpStatus:StartContinuousExportResponse'

-> StartContinuousExportResponse

Create a value of StartContinuousExportResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3Bucket:StartContinuousExportResponse', startContinuousExportResponse_s3Bucket - The name of the s3 bucket where the export data parquet files are stored.

$sel:startTime:StartContinuousExportResponse', startContinuousExportResponse_startTime - The timestamp representing when the continuous export was started.

$sel:dataSource:StartContinuousExportResponse', startContinuousExportResponse_dataSource - The type of data collector used to gather this data (currently only offered for AGENT).

$sel:schemaStorageConfig:StartContinuousExportResponse', startContinuousExportResponse_schemaStorageConfig - A dictionary which describes how the data is stored.

databaseName - the name of the Glue database used to store the schema.

$sel:exportId:StartContinuousExportResponse', startContinuousExportResponse_exportId - The unique ID assigned to this export.

$sel:httpStatus:StartContinuousExportResponse', startContinuousExportResponse_httpStatus - The response's http status code.

Response Lenses

startContinuousExportResponse_s3Bucket :: Lens' StartContinuousExportResponse (Maybe Text)

The name of the s3 bucket where the export data parquet files are stored.

startContinuousExportResponse_startTime :: Lens' StartContinuousExportResponse (Maybe UTCTime)

The timestamp representing when the continuous export was started.

startContinuousExportResponse_dataSource :: Lens' StartContinuousExportResponse (Maybe DataSource)

The type of data collector used to gather this data (currently only offered for AGENT).

startContinuousExportResponse_schemaStorageConfig :: Lens' StartContinuousExportResponse (Maybe (HashMap Text Text))

A dictionary which describes how the data is stored.

databaseName - the name of the Glue database used to store the schema.

startContinuousExportResponse_exportId :: Lens' StartContinuousExportResponse (Maybe Text)

The unique ID assigned to this export.

startContinuousExportResponse_httpStatus :: Lens' StartContinuousExportResponse Int

The response's http status code.