Network.AWS.Discovery.Types

Description

Service Configuration

defaultService :: Service

API version 2015-11-01 of the Amazon Application Discovery Service SDK configuration.

Errors

_ConflictErrorException :: AsError a => Getting (First ServiceError) a ServiceError

_OperationNotPermittedException :: AsError a => Getting (First ServiceError) a ServiceError

This operation is not permitted.

_ServerInternalErrorException :: AsError a => Getting (First ServiceError) a ServiceError

The server experienced an internal error. Try again.

_InvalidParameterException :: AsError a => Getting (First ServiceError) a ServiceError

One or more parameters are not valid. Verify the parameters and try again.

_HomeRegionNotSetException :: AsError a => Getting (First ServiceError) a ServiceError

The home region is not set. Set the home region to continue.

_AuthorizationErrorException :: AsError a => Getting (First ServiceError) a ServiceError

The AWS user account does not have permission to perform the action. Check the IAM policy associated with this account.

_InvalidParameterValueException :: AsError a => Getting (First ServiceError) a ServiceError

The value of one or more parameters are either invalid or out of range. Verify the parameter values and try again.

_ResourceInUseException :: AsError a => Getting (First ServiceError) a ServiceError

This issue occurs when the same clientRequestToken is used with the StartImportTask action, but with different parameters. For example, you use the same request token but have two different import URLs, you can encounter this issue. If the import tasks are meant to be different, use a different clientRequestToken, and try again.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified configuration ID was not located. Verify the configuration ID and try again.

AgentStatus

newtype AgentStatus

Constructors

AgentStatus'

Fields

fromAgentStatus :: Text

Bundled Patterns

pattern AgentStatus_BLACKLISTED :: AgentStatus

pattern AgentStatus_HEALTHY :: AgentStatus

pattern AgentStatus_RUNNING :: AgentStatus

pattern AgentStatus_SHUTDOWN :: AgentStatus

pattern AgentStatus_UNHEALTHY :: AgentStatus

pattern AgentStatus_UNKNOWN :: AgentStatus

BatchDeleteImportDataErrorCode

newtype BatchDeleteImportDataErrorCode

Constructors

BatchDeleteImportDataErrorCode'

Fields

fromBatchDeleteImportDataErrorCode :: Text

Bundled Patterns

pattern BatchDeleteImportDataErrorCode_INTERNAL_SERVER_ERROR :: BatchDeleteImportDataErrorCode

pattern BatchDeleteImportDataErrorCode_NOT_FOUND :: BatchDeleteImportDataErrorCode

pattern BatchDeleteImportDataErrorCode_OVER_LIMIT :: BatchDeleteImportDataErrorCode

ConfigurationItemType

newtype ConfigurationItemType

Constructors

ConfigurationItemType'

Fields

fromConfigurationItemType :: Text

Bundled Patterns

pattern ConfigurationItemType_APPLICATION :: ConfigurationItemType

pattern ConfigurationItemType_CONNECTION :: ConfigurationItemType

pattern ConfigurationItemType_PROCESS :: ConfigurationItemType

pattern ConfigurationItemType_SERVER :: ConfigurationItemType

ContinuousExportStatus

newtype ContinuousExportStatus

Constructors

ContinuousExportStatus'

Fields

fromContinuousExportStatus :: Text

Bundled Patterns

pattern ContinuousExportStatus_ACTIVE :: ContinuousExportStatus

pattern ContinuousExportStatus_ERROR :: ContinuousExportStatus

pattern ContinuousExportStatus_INACTIVE :: ContinuousExportStatus

pattern ContinuousExportStatus_START_FAILED :: ContinuousExportStatus

pattern ContinuousExportStatus_START_IN_PROGRESS :: ContinuousExportStatus

pattern ContinuousExportStatus_STOP_FAILED :: ContinuousExportStatus

pattern ContinuousExportStatus_STOP_IN_PROGRESS :: ContinuousExportStatus

DataSource

newtype DataSource

Constructors

DataSource'

Fields

fromDataSource :: Text

Bundled Patterns

pattern DataSource_AGENT :: DataSource

ExportDataFormat

newtype ExportDataFormat

Constructors

ExportDataFormat'

Fields

fromExportDataFormat :: Text

Bundled Patterns

pattern ExportDataFormat_CSV :: ExportDataFormat

pattern ExportDataFormat_GRAPHML :: ExportDataFormat

ExportStatus

newtype ExportStatus

Constructors

ExportStatus'

Fields

fromExportStatus :: Text

Bundled Patterns

pattern ExportStatus_FAILED :: ExportStatus

pattern ExportStatus_IN_PROGRESS :: ExportStatus

pattern ExportStatus_SUCCEEDED :: ExportStatus

ImportStatus

newtype ImportStatus

Constructors

ImportStatus'

Fields

fromImportStatus :: Text

Bundled Patterns

pattern ImportStatus_DELETE_COMPLETE :: ImportStatus

pattern ImportStatus_DELETE_FAILED :: ImportStatus

pattern ImportStatus_DELETE_FAILED_LIMIT_EXCEEDED :: ImportStatus

pattern ImportStatus_DELETE_IN_PROGRESS :: ImportStatus

pattern ImportStatus_IMPORT_COMPLETE :: ImportStatus

pattern ImportStatus_IMPORT_COMPLETE_WITH_ERRORS :: ImportStatus

pattern ImportStatus_IMPORT_FAILED :: ImportStatus

pattern ImportStatus_IMPORT_FAILED_RECORD_LIMIT_EXCEEDED :: ImportStatus

pattern ImportStatus_IMPORT_FAILED_SERVER_LIMIT_EXCEEDED :: ImportStatus

pattern ImportStatus_IMPORT_IN_PROGRESS :: ImportStatus

pattern ImportStatus_INTERNAL_ERROR :: ImportStatus

ImportTaskFilterName

newtype ImportTaskFilterName

Constructors

ImportTaskFilterName'

Fields

fromImportTaskFilterName :: Text

Bundled Patterns

pattern ImportTaskFilterName_IMPORT_TASK_ID :: ImportTaskFilterName

pattern ImportTaskFilterName_NAME :: ImportTaskFilterName

pattern ImportTaskFilterName_STATUS :: ImportTaskFilterName

OrderString

newtype OrderString

Constructors

OrderString'

Fields

fromOrderString :: Text

Bundled Patterns

pattern OrderString_ASC :: OrderString

pattern OrderString_DESC :: OrderString

AgentConfigurationStatus

data AgentConfigurationStatus

Information about agents or connectors that were instructed to start collecting data. Information includes the agent/connector ID, a description of the operation, and whether the agent/connector configuration was updated.

See: newAgentConfigurationStatus smart constructor.

Constructors

AgentConfigurationStatus'

Fields

agentId :: Maybe Text

The agent/connector ID.

operationSucceeded :: Maybe Bool

Information about the status of the StartDataCollection and StopDataCollection operations. The system has recorded the data collection operation. The agent/connector receives this command the next time it polls for a new command.

description :: Maybe Text

A description of the operation performed.

newAgentConfigurationStatus :: AgentConfigurationStatus

Create a value of AgentConfigurationStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:agentId:AgentConfigurationStatus', agentConfigurationStatus_agentId - The agent/connector ID.

$sel:operationSucceeded:AgentConfigurationStatus', agentConfigurationStatus_operationSucceeded - Information about the status of the StartDataCollection and StopDataCollection operations. The system has recorded the data collection operation. The agent/connector receives this command the next time it polls for a new command.

$sel:description:AgentConfigurationStatus', agentConfigurationStatus_description - A description of the operation performed.

agentConfigurationStatus_agentId :: Lens' AgentConfigurationStatus (Maybe Text)

The agent/connector ID.

agentConfigurationStatus_operationSucceeded :: Lens' AgentConfigurationStatus (Maybe Bool)

Information about the status of the StartDataCollection and StopDataCollection operations. The system has recorded the data collection operation. The agent/connector receives this command the next time it polls for a new command.

agentConfigurationStatus_description :: Lens' AgentConfigurationStatus (Maybe Text)

A description of the operation performed.

AgentInfo

data AgentInfo

Information about agents or connectors associated with the user’s AWS account. Information includes agent or connector IDs, IP addresses, media access control (MAC) addresses, agent or connector health, hostname where the agent or connector resides, and agent version for each agent.

See: newAgentInfo smart constructor.

Constructors

AgentInfo'

Fields

hostName :: Maybe Text

The name of the host where the agent or connector resides. The host can be a server or virtual machine.

agentId :: Maybe Text

The agent or connector ID.

agentType :: Maybe Text

Type of agent.

connectorId :: Maybe Text

The ID of the connector.

lastHealthPingTime :: Maybe Text

Time since agent or connector health was reported.

agentNetworkInfoList :: Maybe [AgentNetworkInfo]

Network details about the host where the agent or connector resides.

registeredTime :: Maybe Text

Agent's first registration timestamp in UTC.

version :: Maybe Text

The agent or connector version.

health :: Maybe AgentStatus

The health of the agent or connector.

collectionStatus :: Maybe Text

Status of the collection process for an agent or connector.

newAgentInfo :: AgentInfo

Create a value of AgentInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hostName:AgentInfo', agentInfo_hostName - The name of the host where the agent or connector resides. The host can be a server or virtual machine.

$sel:agentId:AgentInfo', agentInfo_agentId - The agent or connector ID.

$sel:agentType:AgentInfo', agentInfo_agentType - Type of agent.

$sel:connectorId:AgentInfo', agentInfo_connectorId - The ID of the connector.

$sel:lastHealthPingTime:AgentInfo', agentInfo_lastHealthPingTime - Time since agent or connector health was reported.

$sel:agentNetworkInfoList:AgentInfo', agentInfo_agentNetworkInfoList - Network details about the host where the agent or connector resides.

$sel:registeredTime:AgentInfo', agentInfo_registeredTime - Agent's first registration timestamp in UTC.

$sel:version:AgentInfo', agentInfo_version - The agent or connector version.

$sel:health:AgentInfo', agentInfo_health - The health of the agent or connector.

$sel:collectionStatus:AgentInfo', agentInfo_collectionStatus - Status of the collection process for an agent or connector.

agentInfo_hostName :: Lens' AgentInfo (Maybe Text)

The name of the host where the agent or connector resides. The host can be a server or virtual machine.

agentInfo_agentId :: Lens' AgentInfo (Maybe Text)

The agent or connector ID.

agentInfo_agentType :: Lens' AgentInfo (Maybe Text)

Type of agent.

agentInfo_connectorId :: Lens' AgentInfo (Maybe Text)

The ID of the connector.

agentInfo_lastHealthPingTime :: Lens' AgentInfo (Maybe Text)

Time since agent or connector health was reported.

agentInfo_agentNetworkInfoList :: Lens' AgentInfo (Maybe [AgentNetworkInfo])

Network details about the host where the agent or connector resides.

agentInfo_registeredTime :: Lens' AgentInfo (Maybe Text)

Agent's first registration timestamp in UTC.

agentInfo_version :: Lens' AgentInfo (Maybe Text)

The agent or connector version.

agentInfo_health :: Lens' AgentInfo (Maybe AgentStatus)

The health of the agent or connector.

agentInfo_collectionStatus :: Lens' AgentInfo (Maybe Text)

Status of the collection process for an agent or connector.

AgentNetworkInfo

data AgentNetworkInfo

Network details about the host where the agent/connector resides.

See: newAgentNetworkInfo smart constructor.

Constructors

AgentNetworkInfo'

Fields

macAddress :: Maybe Text

The MAC address for the host where the agent/connector resides.

ipAddress :: Maybe Text

The IP address for the host where the agent/connector resides.

newAgentNetworkInfo :: AgentNetworkInfo

Create a value of AgentNetworkInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:macAddress:AgentNetworkInfo', agentNetworkInfo_macAddress - The MAC address for the host where the agent/connector resides.

$sel:ipAddress:AgentNetworkInfo', agentNetworkInfo_ipAddress - The IP address for the host where the agent/connector resides.

agentNetworkInfo_macAddress :: Lens' AgentNetworkInfo (Maybe Text)

The MAC address for the host where the agent/connector resides.

agentNetworkInfo_ipAddress :: Lens' AgentNetworkInfo (Maybe Text)

The IP address for the host where the agent/connector resides.

BatchDeleteImportDataError

data BatchDeleteImportDataError

Error messages returned for each import task that you deleted as a response for this command.

See: newBatchDeleteImportDataError smart constructor.

Constructors

BatchDeleteImportDataError'

Fields

errorDescription :: Maybe Text

The description of the error that occurred for a specific import task.

importTaskId :: Maybe Text

The unique import ID associated with the error that occurred.

errorCode :: Maybe BatchDeleteImportDataErrorCode

The type of error that occurred for a specific import task.

newBatchDeleteImportDataError :: BatchDeleteImportDataError

Create a value of BatchDeleteImportDataError with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:errorDescription:BatchDeleteImportDataError', batchDeleteImportDataError_errorDescription - The description of the error that occurred for a specific import task.

$sel:importTaskId:BatchDeleteImportDataError', batchDeleteImportDataError_importTaskId - The unique import ID associated with the error that occurred.

$sel:errorCode:BatchDeleteImportDataError', batchDeleteImportDataError_errorCode - The type of error that occurred for a specific import task.

batchDeleteImportDataError_errorDescription :: Lens' BatchDeleteImportDataError (Maybe Text)

The description of the error that occurred for a specific import task.

batchDeleteImportDataError_importTaskId :: Lens' BatchDeleteImportDataError (Maybe Text)

The unique import ID associated with the error that occurred.

batchDeleteImportDataError_errorCode :: Lens' BatchDeleteImportDataError (Maybe BatchDeleteImportDataErrorCode)

The type of error that occurred for a specific import task.

ConfigurationTag

data ConfigurationTag

Tags for a configuration item. Tags are metadata that help you categorize IT assets.

See: newConfigurationTag smart constructor.

Constructors

ConfigurationTag'

Fields

key :: Maybe Text

A type of tag on which to filter. For example, serverType.

configurationId :: Maybe Text

The configuration ID for the item to tag. You can specify a list of keys and values.

value :: Maybe Text

A value on which to filter. For example key = serverType and /value = web server/.

configurationType :: Maybe ConfigurationItemType

A type of IT asset to tag.

timeOfCreation :: Maybe POSIX

The time the configuration tag was created in Coordinated Universal Time (UTC).

newConfigurationTag :: ConfigurationTag

Create a value of ConfigurationTag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:ConfigurationTag', configurationTag_key - A type of tag on which to filter. For example, serverType.

$sel:configurationId:ConfigurationTag', configurationTag_configurationId - The configuration ID for the item to tag. You can specify a list of keys and values.

$sel:value:ConfigurationTag', configurationTag_value - A value on which to filter. For example key = serverType and /value = web server/.

$sel:configurationType:ConfigurationTag', configurationTag_configurationType - A type of IT asset to tag.

$sel:timeOfCreation:ConfigurationTag', configurationTag_timeOfCreation - The time the configuration tag was created in Coordinated Universal Time (UTC).

configurationTag_key :: Lens' ConfigurationTag (Maybe Text)

A type of tag on which to filter. For example, serverType.

configurationTag_configurationId :: Lens' ConfigurationTag (Maybe Text)

The configuration ID for the item to tag. You can specify a list of keys and values.

configurationTag_value :: Lens' ConfigurationTag (Maybe Text)

A value on which to filter. For example key = serverType and /value = web server/.

configurationTag_configurationType :: Lens' ConfigurationTag (Maybe ConfigurationItemType)

A type of IT asset to tag.

configurationTag_timeOfCreation :: Lens' ConfigurationTag (Maybe UTCTime)

The time the configuration tag was created in Coordinated Universal Time (UTC).

ContinuousExportDescription

data ContinuousExportDescription

A list of continuous export descriptions.

See: newContinuousExportDescription smart constructor.

Constructors

ContinuousExportDescription'

Fields

status :: Maybe ContinuousExportStatus

Describes the status of the export. Can be one of the following values:

START_IN_PROGRESS - setting up resources to start continuous export.

START_FAILED - an error occurred setting up continuous export. To recover, call start-continuous-export again.

ACTIVE - data is being exported to the customer bucket.

ERROR - an error occurred during export. To fix the issue, call stop-continuous-export and start-continuous-export.

STOP_IN_PROGRESS - stopping the export.

STOP_FAILED - an error occurred stopping the export. To recover, call stop-continuous-export again.

INACTIVE - the continuous export has been stopped. Data is no longer being exported to the customer bucket.

s3Bucket :: Maybe Text

The name of the s3 bucket where the export data parquet files are stored.

stopTime :: Maybe POSIX

The timestamp that represents when this continuous export was stopped.

startTime :: Maybe POSIX

The timestamp representing when the continuous export was started.

dataSource :: Maybe DataSource

The type of data collector used to gather this data (currently only offered for AGENT).

statusDetail :: Maybe Text

Contains information about any errors that have occurred. This data type can have the following values:

ACCESS_DENIED - You don’t have permission to start Data Exploration in Amazon Athena. Contact your AWS administrator for help. For more information, see Setting Up AWS Application Discovery Service in the Application Discovery Service User Guide.

DELIVERY_STREAM_LIMIT_FAILURE - You reached the limit for Amazon Kinesis Data Firehose delivery streams. Reduce the number of streams or request a limit increase and try again. For more information, see Kinesis Data Streams Limits in the Amazon Kinesis Data Streams Developer Guide.

FIREHOSE_ROLE_MISSING - The Data Exploration feature is in an error state because your IAM User is missing the AWSApplicationDiscoveryServiceFirehose role. Turn on Data Exploration in Amazon Athena and try again. For more information, see Step 3: Provide Application Discovery Service Access to Non-Administrator Users by Attaching Policies in the Application Discovery Service User Guide.

FIREHOSE_STREAM_DOES_NOT_EXIST - The Data Exploration feature is in an error state because your IAM User is missing one or more of the Kinesis data delivery streams.

INTERNAL_FAILURE - The Data Exploration feature is in an error state because of an internal failure. Try again later. If this problem persists, contact AWS Support.

S3_BUCKET_LIMIT_FAILURE - You reached the limit for Amazon S3 buckets. Reduce the number of Amazon S3 buckets or request a limit increase and try again. For more information, see Bucket Restrictions and Limitations in the Amazon Simple Storage Service Developer Guide.

S3_NOT_SIGNED_UP - Your account is not signed up for the Amazon S3 service. You must sign up before you can use Amazon S3. You can sign up at the following URL: https://aws.amazon.com/s3.

schemaStorageConfig :: Maybe (HashMap Text Text)

An object which describes how the data is stored.

databaseName - the name of the Glue database used to store the schema.

exportId :: Maybe Text

The unique ID assigned to this export.

newContinuousExportDescription :: ContinuousExportDescription

Create a value of ContinuousExportDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ContinuousExportDescription', continuousExportDescription_status - Describes the status of the export. Can be one of the following values:

START_IN_PROGRESS - setting up resources to start continuous export.

START_FAILED - an error occurred setting up continuous export. To recover, call start-continuous-export again.

ACTIVE - data is being exported to the customer bucket.

ERROR - an error occurred during export. To fix the issue, call stop-continuous-export and start-continuous-export.

STOP_IN_PROGRESS - stopping the export.

STOP_FAILED - an error occurred stopping the export. To recover, call stop-continuous-export again.

INACTIVE - the continuous export has been stopped. Data is no longer being exported to the customer bucket.

$sel:s3Bucket:ContinuousExportDescription', continuousExportDescription_s3Bucket - The name of the s3 bucket where the export data parquet files are stored.

$sel:stopTime:ContinuousExportDescription', continuousExportDescription_stopTime - The timestamp that represents when this continuous export was stopped.

$sel:startTime:ContinuousExportDescription', continuousExportDescription_startTime - The timestamp representing when the continuous export was started.

$sel:dataSource:ContinuousExportDescription', continuousExportDescription_dataSource - The type of data collector used to gather this data (currently only offered for AGENT).

$sel:statusDetail:ContinuousExportDescription', continuousExportDescription_statusDetail - Contains information about any errors that have occurred. This data type can have the following values:

ACCESS_DENIED - You don’t have permission to start Data Exploration in Amazon Athena. Contact your AWS administrator for help. For more information, see Setting Up AWS Application Discovery Service in the Application Discovery Service User Guide.

DELIVERY_STREAM_LIMIT_FAILURE - You reached the limit for Amazon Kinesis Data Firehose delivery streams. Reduce the number of streams or request a limit increase and try again. For more information, see Kinesis Data Streams Limits in the Amazon Kinesis Data Streams Developer Guide.

FIREHOSE_ROLE_MISSING - The Data Exploration feature is in an error state because your IAM User is missing the AWSApplicationDiscoveryServiceFirehose role. Turn on Data Exploration in Amazon Athena and try again. For more information, see Step 3: Provide Application Discovery Service Access to Non-Administrator Users by Attaching Policies in the Application Discovery Service User Guide.

FIREHOSE_STREAM_DOES_NOT_EXIST - The Data Exploration feature is in an error state because your IAM User is missing one or more of the Kinesis data delivery streams.

INTERNAL_FAILURE - The Data Exploration feature is in an error state because of an internal failure. Try again later. If this problem persists, contact AWS Support.

S3_BUCKET_LIMIT_FAILURE - You reached the limit for Amazon S3 buckets. Reduce the number of Amazon S3 buckets or request a limit increase and try again. For more information, see Bucket Restrictions and Limitations in the Amazon Simple Storage Service Developer Guide.

S3_NOT_SIGNED_UP - Your account is not signed up for the Amazon S3 service. You must sign up before you can use Amazon S3. You can sign up at the following URL: https://aws.amazon.com/s3.

$sel:schemaStorageConfig:ContinuousExportDescription', continuousExportDescription_schemaStorageConfig - An object which describes how the data is stored.

databaseName - the name of the Glue database used to store the schema.

$sel:exportId:ContinuousExportDescription', continuousExportDescription_exportId - The unique ID assigned to this export.

continuousExportDescription_status :: Lens' ContinuousExportDescription (Maybe ContinuousExportStatus)

Describes the status of the export. Can be one of the following values:

START_IN_PROGRESS - setting up resources to start continuous export.

START_FAILED - an error occurred setting up continuous export. To recover, call start-continuous-export again.

ACTIVE - data is being exported to the customer bucket.

ERROR - an error occurred during export. To fix the issue, call stop-continuous-export and start-continuous-export.

STOP_IN_PROGRESS - stopping the export.

STOP_FAILED - an error occurred stopping the export. To recover, call stop-continuous-export again.

INACTIVE - the continuous export has been stopped. Data is no longer being exported to the customer bucket.

continuousExportDescription_s3Bucket :: Lens' ContinuousExportDescription (Maybe Text)

The name of the s3 bucket where the export data parquet files are stored.

continuousExportDescription_stopTime :: Lens' ContinuousExportDescription (Maybe UTCTime)

The timestamp that represents when this continuous export was stopped.

continuousExportDescription_startTime :: Lens' ContinuousExportDescription (Maybe UTCTime)

The timestamp representing when the continuous export was started.

continuousExportDescription_dataSource :: Lens' ContinuousExportDescription (Maybe DataSource)

The type of data collector used to gather this data (currently only offered for AGENT).

continuousExportDescription_statusDetail :: Lens' ContinuousExportDescription (Maybe Text)

Contains information about any errors that have occurred. This data type can have the following values:

ACCESS_DENIED - You don’t have permission to start Data Exploration in Amazon Athena. Contact your AWS administrator for help. For more information, see Setting Up AWS Application Discovery Service in the Application Discovery Service User Guide.

DELIVERY_STREAM_LIMIT_FAILURE - You reached the limit for Amazon Kinesis Data Firehose delivery streams. Reduce the number of streams or request a limit increase and try again. For more information, see Kinesis Data Streams Limits in the Amazon Kinesis Data Streams Developer Guide.

FIREHOSE_ROLE_MISSING - The Data Exploration feature is in an error state because your IAM User is missing the AWSApplicationDiscoveryServiceFirehose role. Turn on Data Exploration in Amazon Athena and try again. For more information, see Step 3: Provide Application Discovery Service Access to Non-Administrator Users by Attaching Policies in the Application Discovery Service User Guide.

FIREHOSE_STREAM_DOES_NOT_EXIST - The Data Exploration feature is in an error state because your IAM User is missing one or more of the Kinesis data delivery streams.

INTERNAL_FAILURE - The Data Exploration feature is in an error state because of an internal failure. Try again later. If this problem persists, contact AWS Support.

S3_BUCKET_LIMIT_FAILURE - You reached the limit for Amazon S3 buckets. Reduce the number of Amazon S3 buckets or request a limit increase and try again. For more information, see Bucket Restrictions and Limitations in the Amazon Simple Storage Service Developer Guide.

S3_NOT_SIGNED_UP - Your account is not signed up for the Amazon S3 service. You must sign up before you can use Amazon S3. You can sign up at the following URL: https://aws.amazon.com/s3.

continuousExportDescription_schemaStorageConfig :: Lens' ContinuousExportDescription (Maybe (HashMap Text Text))

An object which describes how the data is stored.

databaseName - the name of the Glue database used to store the schema.

continuousExportDescription_exportId :: Lens' ContinuousExportDescription (Maybe Text)

The unique ID assigned to this export.

CustomerAgentInfo

data CustomerAgentInfo

Inventory data for installed discovery agents.

See: newCustomerAgentInfo smart constructor.

Constructors

CustomerAgentInfo'

Fields

activeAgents :: Int

Number of active discovery agents.

healthyAgents :: Int

Number of healthy discovery agents

blackListedAgents :: Int

Number of blacklisted discovery agents.

shutdownAgents :: Int

Number of discovery agents with status SHUTDOWN.

unhealthyAgents :: Int

Number of unhealthy discovery agents.

totalAgents :: Int

Total number of discovery agents.

unknownAgents :: Int

Number of unknown discovery agents.

newCustomerAgentInfo

Arguments

:: Int
$sel:activeAgents:CustomerAgentInfo'

-> Int
$sel:healthyAgents:CustomerAgentInfo'

-> Int
$sel:blackListedAgents:CustomerAgentInfo'

-> Int
$sel:shutdownAgents:CustomerAgentInfo'

-> Int
$sel:unhealthyAgents:CustomerAgentInfo'

-> Int
$sel:totalAgents:CustomerAgentInfo'

-> Int
$sel:unknownAgents:CustomerAgentInfo'

-> CustomerAgentInfo

Create a value of CustomerAgentInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:activeAgents:CustomerAgentInfo', customerAgentInfo_activeAgents - Number of active discovery agents.

$sel:healthyAgents:CustomerAgentInfo', customerAgentInfo_healthyAgents - Number of healthy discovery agents

$sel:blackListedAgents:CustomerAgentInfo', customerAgentInfo_blackListedAgents - Number of blacklisted discovery agents.

$sel:shutdownAgents:CustomerAgentInfo', customerAgentInfo_shutdownAgents - Number of discovery agents with status SHUTDOWN.

$sel:unhealthyAgents:CustomerAgentInfo', customerAgentInfo_unhealthyAgents - Number of unhealthy discovery agents.

$sel:totalAgents:CustomerAgentInfo', customerAgentInfo_totalAgents - Total number of discovery agents.

$sel:unknownAgents:CustomerAgentInfo', customerAgentInfo_unknownAgents - Number of unknown discovery agents.

customerAgentInfo_activeAgents :: Lens' CustomerAgentInfo Int

Number of active discovery agents.

customerAgentInfo_healthyAgents :: Lens' CustomerAgentInfo Int

Number of healthy discovery agents

customerAgentInfo_blackListedAgents :: Lens' CustomerAgentInfo Int

Number of blacklisted discovery agents.

customerAgentInfo_shutdownAgents :: Lens' CustomerAgentInfo Int

Number of discovery agents with status SHUTDOWN.

customerAgentInfo_unhealthyAgents :: Lens' CustomerAgentInfo Int

Number of unhealthy discovery agents.

customerAgentInfo_totalAgents :: Lens' CustomerAgentInfo Int

Total number of discovery agents.

customerAgentInfo_unknownAgents :: Lens' CustomerAgentInfo Int

Number of unknown discovery agents.

CustomerConnectorInfo

data CustomerConnectorInfo

Inventory data for installed discovery connectors.

See: newCustomerConnectorInfo smart constructor.

Constructors

CustomerConnectorInfo'

Fields

activeConnectors :: Int

Number of active discovery connectors.

healthyConnectors :: Int

Number of healthy discovery connectors.

blackListedConnectors :: Int

Number of blacklisted discovery connectors.

shutdownConnectors :: Int

Number of discovery connectors with status SHUTDOWN,

unhealthyConnectors :: Int

Number of unhealthy discovery connectors.

totalConnectors :: Int

Total number of discovery connectors.

unknownConnectors :: Int

Number of unknown discovery connectors.

newCustomerConnectorInfo

Arguments

:: Int
$sel:activeConnectors:CustomerConnectorInfo'

-> Int
$sel:healthyConnectors:CustomerConnectorInfo'

-> Int
$sel:blackListedConnectors:CustomerConnectorInfo'

-> Int
$sel:shutdownConnectors:CustomerConnectorInfo'

-> Int
$sel:unhealthyConnectors:CustomerConnectorInfo'

-> Int
$sel:totalConnectors:CustomerConnectorInfo'

-> Int
$sel:unknownConnectors:CustomerConnectorInfo'

-> CustomerConnectorInfo

Create a value of CustomerConnectorInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:activeConnectors:CustomerConnectorInfo', customerConnectorInfo_activeConnectors - Number of active discovery connectors.

$sel:healthyConnectors:CustomerConnectorInfo', customerConnectorInfo_healthyConnectors - Number of healthy discovery connectors.

$sel:blackListedConnectors:CustomerConnectorInfo', customerConnectorInfo_blackListedConnectors - Number of blacklisted discovery connectors.

$sel:shutdownConnectors:CustomerConnectorInfo', customerConnectorInfo_shutdownConnectors - Number of discovery connectors with status SHUTDOWN,

$sel:unhealthyConnectors:CustomerConnectorInfo', customerConnectorInfo_unhealthyConnectors - Number of unhealthy discovery connectors.

$sel:totalConnectors:CustomerConnectorInfo', customerConnectorInfo_totalConnectors - Total number of discovery connectors.

$sel:unknownConnectors:CustomerConnectorInfo', customerConnectorInfo_unknownConnectors - Number of unknown discovery connectors.

customerConnectorInfo_activeConnectors :: Lens' CustomerConnectorInfo Int

Number of active discovery connectors.

customerConnectorInfo_healthyConnectors :: Lens' CustomerConnectorInfo Int

Number of healthy discovery connectors.

customerConnectorInfo_blackListedConnectors :: Lens' CustomerConnectorInfo Int

Number of blacklisted discovery connectors.

customerConnectorInfo_shutdownConnectors :: Lens' CustomerConnectorInfo Int

Number of discovery connectors with status SHUTDOWN,

customerConnectorInfo_unhealthyConnectors :: Lens' CustomerConnectorInfo Int

Number of unhealthy discovery connectors.

customerConnectorInfo_totalConnectors :: Lens' CustomerConnectorInfo Int

Total number of discovery connectors.

customerConnectorInfo_unknownConnectors :: Lens' CustomerConnectorInfo Int

Number of unknown discovery connectors.

ExportFilter

data ExportFilter

Used to select which agent's data is to be exported. A single agent ID may be selected for export using the StartExportTask action.

See: newExportFilter smart constructor.

Constructors

ExportFilter'

Fields

name :: Text

A single ExportFilter name. Supported filters: agentId.

values :: [Text]

A single agentId for a Discovery Agent. An agentId can be found using the DescribeAgents action. Typically an ADS agentId is in the form o-0123456789abcdef0.

condition :: Text

Supported condition: EQUALS

newExportFilter

Arguments

:: Text
$sel:name:ExportFilter'

-> Text
$sel:condition:ExportFilter'

-> ExportFilter

Create a value of ExportFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:ExportFilter', exportFilter_name - A single ExportFilter name. Supported filters: agentId.

$sel:values:ExportFilter', exportFilter_values - A single agentId for a Discovery Agent. An agentId can be found using the DescribeAgents action. Typically an ADS agentId is in the form o-0123456789abcdef0.

$sel:condition:ExportFilter', exportFilter_condition - Supported condition: EQUALS

exportFilter_name :: Lens' ExportFilter Text

A single ExportFilter name. Supported filters: agentId.

exportFilter_values :: Lens' ExportFilter [Text]

A single agentId for a Discovery Agent. An agentId can be found using the DescribeAgents action. Typically an ADS agentId is in the form o-0123456789abcdef0.

exportFilter_condition :: Lens' ExportFilter Text

Supported condition: EQUALS

ExportInfo

data ExportInfo

Information regarding the export status of discovered data. The value is an array of objects.

See: newExportInfo smart constructor.

Constructors

ExportInfo'

Fields

isTruncated :: Maybe Bool

If true, the export of agent information exceeded the size limit for a single export and the exported data is incomplete for the requested time range. To address this, select a smaller time range for the export by using startDate and endDate.

configurationsDownloadUrl :: Maybe Text

A URL for an Amazon S3 bucket where you can review the exported data. The URL is displayed only if the export succeeded.

requestedStartTime :: Maybe POSIX

The value of startTime parameter in the StartExportTask request. If no startTime was requested, this result does not appear in ExportInfo.

requestedEndTime :: Maybe POSIX

The endTime used in the StartExportTask request. If no endTime was requested, this result does not appear in ExportInfo.

exportId :: Text

A unique identifier used to query an export.

exportStatus :: ExportStatus

The status of the data export job.

statusMessage :: Text

A status message provided for API callers.

exportRequestTime :: POSIX

The time that the data export was initiated.

newExportInfo

Arguments

:: Text
$sel:exportId:ExportInfo'

-> ExportStatus
$sel:exportStatus:ExportInfo'

-> Text
$sel:statusMessage:ExportInfo'

-> UTCTime
$sel:exportRequestTime:ExportInfo'

-> ExportInfo

Create a value of ExportInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:isTruncated:ExportInfo', exportInfo_isTruncated - If true, the export of agent information exceeded the size limit for a single export and the exported data is incomplete for the requested time range. To address this, select a smaller time range for the export by using startDate and endDate.

$sel:configurationsDownloadUrl:ExportInfo', exportInfo_configurationsDownloadUrl - A URL for an Amazon S3 bucket where you can review the exported data. The URL is displayed only if the export succeeded.

$sel:requestedStartTime:ExportInfo', exportInfo_requestedStartTime - The value of startTime parameter in the StartExportTask request. If no startTime was requested, this result does not appear in ExportInfo.

$sel:requestedEndTime:ExportInfo', exportInfo_requestedEndTime - The endTime used in the StartExportTask request. If no endTime was requested, this result does not appear in ExportInfo.

$sel:exportId:ExportInfo', exportInfo_exportId - A unique identifier used to query an export.

$sel:exportStatus:ExportInfo', exportInfo_exportStatus - The status of the data export job.

$sel:statusMessage:ExportInfo', exportInfo_statusMessage - A status message provided for API callers.

$sel:exportRequestTime:ExportInfo', exportInfo_exportRequestTime - The time that the data export was initiated.

exportInfo_isTruncated :: Lens' ExportInfo (Maybe Bool)

If true, the export of agent information exceeded the size limit for a single export and the exported data is incomplete for the requested time range. To address this, select a smaller time range for the export by using startDate and endDate.

exportInfo_configurationsDownloadUrl :: Lens' ExportInfo (Maybe Text)

A URL for an Amazon S3 bucket where you can review the exported data. The URL is displayed only if the export succeeded.

exportInfo_requestedStartTime :: Lens' ExportInfo (Maybe UTCTime)

The value of startTime parameter in the StartExportTask request. If no startTime was requested, this result does not appear in ExportInfo.

exportInfo_requestedEndTime :: Lens' ExportInfo (Maybe UTCTime)

The endTime used in the StartExportTask request. If no endTime was requested, this result does not appear in ExportInfo.

exportInfo_exportId :: Lens' ExportInfo Text

A unique identifier used to query an export.

exportInfo_exportStatus :: Lens' ExportInfo ExportStatus

The status of the data export job.

exportInfo_statusMessage :: Lens' ExportInfo Text

A status message provided for API callers.

exportInfo_exportRequestTime :: Lens' ExportInfo UTCTime

The time that the data export was initiated.

Filter

data Filter

A filter that can use conditional operators.

For more information about filters, see Querying Discovered Configuration Items in the AWS Application Discovery Service User Guide.

See: newFilter smart constructor.

Constructors

Filter'

Fields

name :: Text

The name of the filter.

values :: [Text]

A string value on which to filter. For example, if you choose the destinationServer.osVersion filter name, you could specify Ubuntu for the value.

condition :: Text

A conditional operator. The following operators are valid: EQUALS, NOT_EQUALS, CONTAINS, NOT_CONTAINS. If you specify multiple filters, the system utilizes all filters as though concatenated by AND. If you specify multiple values for a particular filter, the system differentiates the values using OR. Calling either DescribeConfigurations or ListConfigurations returns attributes of matching configuration items.

newFilter

Arguments

:: Text
$sel:name:Filter'

-> Text
$sel:condition:Filter'

-> Filter

Create a value of Filter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:Filter', filter_name - The name of the filter.

$sel:values:Filter', filter_values - A string value on which to filter. For example, if you choose the destinationServer.osVersion filter name, you could specify Ubuntu for the value.

$sel:condition:Filter', filter_condition - A conditional operator. The following operators are valid: EQUALS, NOT_EQUALS, CONTAINS, NOT_CONTAINS. If you specify multiple filters, the system utilizes all filters as though concatenated by AND. If you specify multiple values for a particular filter, the system differentiates the values using OR. Calling either DescribeConfigurations or ListConfigurations returns attributes of matching configuration items.

filter_name :: Lens' Filter Text

The name of the filter.

filter_values :: Lens' Filter [Text]

A string value on which to filter. For example, if you choose the destinationServer.osVersion filter name, you could specify Ubuntu for the value.

filter_condition :: Lens' Filter Text

A conditional operator. The following operators are valid: EQUALS, NOT_EQUALS, CONTAINS, NOT_CONTAINS. If you specify multiple filters, the system utilizes all filters as though concatenated by AND. If you specify multiple values for a particular filter, the system differentiates the values using OR. Calling either DescribeConfigurations or ListConfigurations returns attributes of matching configuration items.

ImportTask

data ImportTask

An array of information related to the import task request that includes status information, times, IDs, the Amazon S3 Object URL for the import file, and more.

See: newImportTask smart constructor.

Constructors

ImportTask'

Fields

applicationImportSuccess :: Maybe Int

The total number of application records in the import file that were successfully imported.

status :: Maybe ImportStatus

The status of the import task. An import can have the status of IMPORT_COMPLETE and still have some records fail to import from the overall request. More information can be found in the downloadable archive defined in the errorsAndFailedEntriesZip field, or in the Migration Hub management console.

importRequestTime :: Maybe POSIX

The time that the import task request was made, presented in the Unix time stamp format.

serverImportFailure :: Maybe Int

The total number of server records in the import file that failed to be imported.

importTaskId :: Maybe Text

The unique ID for a specific import task. These IDs aren't globally unique, but they are unique within an AWS account.

errorsAndFailedEntriesZip :: Maybe Text

A link to a compressed archive folder (in the ZIP format) that contains an error log and a file of failed records. You can use these two files to quickly identify records that failed, why they failed, and correct those records. Afterward, you can upload the corrected file to your Amazon S3 bucket and create another import task request.

This field also includes authorization information so you can confirm the authenticity of the compressed archive before you download it.

If some records failed to be imported we recommend that you correct the records in the failed entries file and then imports that failed entries file. This prevents you from having to correct and update the larger original file and attempt importing it again.

applicationImportFailure :: Maybe Int

The total number of application records in the import file that failed to be imported.

name :: Maybe Text

A descriptive name for an import task. You can use this name to filter future requests related to this import task, such as identifying applications and servers that were included in this import task. We recommend that you use a meaningful name for each import task.

importCompletionTime :: Maybe POSIX

The time that the import task request finished, presented in the Unix time stamp format.

importUrl :: Maybe Text

The URL for your import file that you've uploaded to Amazon S3.

serverImportSuccess :: Maybe Int

The total number of server records in the import file that were successfully imported.

clientRequestToken :: Maybe Text

A unique token used to prevent the same import request from occurring more than once. If you didn't provide a token, a token was automatically generated when the import task request was sent.

importDeletedTime :: Maybe POSIX

The time that the import task request was deleted, presented in the Unix time stamp format.

newImportTask :: ImportTask

Create a value of ImportTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationImportSuccess:ImportTask', importTask_applicationImportSuccess - The total number of application records in the import file that were successfully imported.

$sel:status:ImportTask', importTask_status - The status of the import task. An import can have the status of IMPORT_COMPLETE and still have some records fail to import from the overall request. More information can be found in the downloadable archive defined in the errorsAndFailedEntriesZip field, or in the Migration Hub management console.

$sel:importRequestTime:ImportTask', importTask_importRequestTime - The time that the import task request was made, presented in the Unix time stamp format.

$sel:serverImportFailure:ImportTask', importTask_serverImportFailure - The total number of server records in the import file that failed to be imported.

$sel:importTaskId:ImportTask', importTask_importTaskId - The unique ID for a specific import task. These IDs aren't globally unique, but they are unique within an AWS account.

$sel:errorsAndFailedEntriesZip:ImportTask', importTask_errorsAndFailedEntriesZip - A link to a compressed archive folder (in the ZIP format) that contains an error log and a file of failed records. You can use these two files to quickly identify records that failed, why they failed, and correct those records. Afterward, you can upload the corrected file to your Amazon S3 bucket and create another import task request.

This field also includes authorization information so you can confirm the authenticity of the compressed archive before you download it.

If some records failed to be imported we recommend that you correct the records in the failed entries file and then imports that failed entries file. This prevents you from having to correct and update the larger original file and attempt importing it again.

$sel:applicationImportFailure:ImportTask', importTask_applicationImportFailure - The total number of application records in the import file that failed to be imported.

$sel:name:ImportTask', importTask_name - A descriptive name for an import task. You can use this name to filter future requests related to this import task, such as identifying applications and servers that were included in this import task. We recommend that you use a meaningful name for each import task.

$sel:importCompletionTime:ImportTask', importTask_importCompletionTime - The time that the import task request finished, presented in the Unix time stamp format.

$sel:importUrl:ImportTask', importTask_importUrl - The URL for your import file that you've uploaded to Amazon S3.

$sel:serverImportSuccess:ImportTask', importTask_serverImportSuccess - The total number of server records in the import file that were successfully imported.

$sel:clientRequestToken:ImportTask', importTask_clientRequestToken - A unique token used to prevent the same import request from occurring more than once. If you didn't provide a token, a token was automatically generated when the import task request was sent.

$sel:importDeletedTime:ImportTask', importTask_importDeletedTime - The time that the import task request was deleted, presented in the Unix time stamp format.

importTask_applicationImportSuccess :: Lens' ImportTask (Maybe Int)

The total number of application records in the import file that were successfully imported.

importTask_status :: Lens' ImportTask (Maybe ImportStatus)

The status of the import task. An import can have the status of IMPORT_COMPLETE and still have some records fail to import from the overall request. More information can be found in the downloadable archive defined in the errorsAndFailedEntriesZip field, or in the Migration Hub management console.

importTask_importRequestTime :: Lens' ImportTask (Maybe UTCTime)

The time that the import task request was made, presented in the Unix time stamp format.

importTask_serverImportFailure :: Lens' ImportTask (Maybe Int)

The total number of server records in the import file that failed to be imported.

importTask_importTaskId :: Lens' ImportTask (Maybe Text)

The unique ID for a specific import task. These IDs aren't globally unique, but they are unique within an AWS account.

importTask_errorsAndFailedEntriesZip :: Lens' ImportTask (Maybe Text)

A link to a compressed archive folder (in the ZIP format) that contains an error log and a file of failed records. You can use these two files to quickly identify records that failed, why they failed, and correct those records. Afterward, you can upload the corrected file to your Amazon S3 bucket and create another import task request.

This field also includes authorization information so you can confirm the authenticity of the compressed archive before you download it.

If some records failed to be imported we recommend that you correct the records in the failed entries file and then imports that failed entries file. This prevents you from having to correct and update the larger original file and attempt importing it again.

importTask_applicationImportFailure :: Lens' ImportTask (Maybe Int)

The total number of application records in the import file that failed to be imported.

importTask_name :: Lens' ImportTask (Maybe Text)

A descriptive name for an import task. You can use this name to filter future requests related to this import task, such as identifying applications and servers that were included in this import task. We recommend that you use a meaningful name for each import task.

importTask_importCompletionTime :: Lens' ImportTask (Maybe UTCTime)

The time that the import task request finished, presented in the Unix time stamp format.

importTask_importUrl :: Lens' ImportTask (Maybe Text)

The URL for your import file that you've uploaded to Amazon S3.

importTask_serverImportSuccess :: Lens' ImportTask (Maybe Int)

The total number of server records in the import file that were successfully imported.

importTask_clientRequestToken :: Lens' ImportTask (Maybe Text)

A unique token used to prevent the same import request from occurring more than once. If you didn't provide a token, a token was automatically generated when the import task request was sent.

importTask_importDeletedTime :: Lens' ImportTask (Maybe UTCTime)

The time that the import task request was deleted, presented in the Unix time stamp format.

ImportTaskFilter

data ImportTaskFilter

A name-values pair of elements you can use to filter the results when querying your import tasks. Currently, wildcards are not supported for filters.

When filtering by import status, all other filter values are ignored.

See: newImportTaskFilter smart constructor.

Constructors

ImportTaskFilter'

Fields

values :: Maybe (NonEmpty Text)

An array of strings that you can provide to match against a specific name, status, or import task ID to filter the results for your import task queries.

name :: Maybe ImportTaskFilterName

The name, status, or import task ID for a specific import task.

newImportTaskFilter :: ImportTaskFilter

Create a value of ImportTaskFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:values:ImportTaskFilter', importTaskFilter_values - An array of strings that you can provide to match against a specific name, status, or import task ID to filter the results for your import task queries.

$sel:name:ImportTaskFilter', importTaskFilter_name - The name, status, or import task ID for a specific import task.

importTaskFilter_values :: Lens' ImportTaskFilter (Maybe (NonEmpty Text))

An array of strings that you can provide to match against a specific name, status, or import task ID to filter the results for your import task queries.

importTaskFilter_name :: Lens' ImportTaskFilter (Maybe ImportTaskFilterName)

The name, status, or import task ID for a specific import task.

NeighborConnectionDetail

data NeighborConnectionDetail

Details about neighboring servers.

See: newNeighborConnectionDetail smart constructor.

Constructors

NeighborConnectionDetail'

Fields

transportProtocol :: Maybe Text

The network protocol used for the connection.

destinationPort :: Maybe Int

The destination network port for the connection.

sourceServerId :: Text

The ID of the server that opened the network connection.

destinationServerId :: Text

The ID of the server that accepted the network connection.

connectionsCount :: Integer

The number of open network connections with the neighboring server.

newNeighborConnectionDetail

Arguments

:: Text
$sel:sourceServerId:NeighborConnectionDetail'

-> Text
$sel:destinationServerId:NeighborConnectionDetail'

-> Integer
$sel:connectionsCount:NeighborConnectionDetail'

-> NeighborConnectionDetail

Create a value of NeighborConnectionDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:transportProtocol:NeighborConnectionDetail', neighborConnectionDetail_transportProtocol - The network protocol used for the connection.

$sel:destinationPort:NeighborConnectionDetail', neighborConnectionDetail_destinationPort - The destination network port for the connection.

$sel:sourceServerId:NeighborConnectionDetail', neighborConnectionDetail_sourceServerId - The ID of the server that opened the network connection.

$sel:destinationServerId:NeighborConnectionDetail', neighborConnectionDetail_destinationServerId - The ID of the server that accepted the network connection.

$sel:connectionsCount:NeighborConnectionDetail', neighborConnectionDetail_connectionsCount - The number of open network connections with the neighboring server.

neighborConnectionDetail_transportProtocol :: Lens' NeighborConnectionDetail (Maybe Text)

The network protocol used for the connection.

neighborConnectionDetail_destinationPort :: Lens' NeighborConnectionDetail (Maybe Int)

The destination network port for the connection.

neighborConnectionDetail_sourceServerId :: Lens' NeighborConnectionDetail Text

The ID of the server that opened the network connection.

neighborConnectionDetail_destinationServerId :: Lens' NeighborConnectionDetail Text

The ID of the server that accepted the network connection.

neighborConnectionDetail_connectionsCount :: Lens' NeighborConnectionDetail Integer

The number of open network connections with the neighboring server.

OrderByElement

data OrderByElement

A field and direction for ordered output.

See: newOrderByElement smart constructor.

Constructors

OrderByElement'

Fields

sortOrder :: Maybe OrderString

Ordering direction.

fieldName :: Text

The field on which to order.

newOrderByElement

Arguments

:: Text
$sel:fieldName:OrderByElement'

-> OrderByElement

Create a value of OrderByElement with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sortOrder:OrderByElement', orderByElement_sortOrder - Ordering direction.

$sel:fieldName:OrderByElement', orderByElement_fieldName - The field on which to order.

orderByElement_sortOrder :: Lens' OrderByElement (Maybe OrderString)

Ordering direction.

orderByElement_fieldName :: Lens' OrderByElement Text

The field on which to order.

Tag

data Tag

Metadata that help you categorize IT assets.

See: newTag smart constructor.

Constructors

Tag'

Fields

key :: Text

The type of tag on which to filter.

value :: Text

A value for a tag key on which to filter.

newTag

Arguments

:: Text
$sel:key:Tag'

-> Text
$sel:value:Tag'

-> Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:Tag', tag_key - The type of tag on which to filter.

$sel:value:Tag', tag_value - A value for a tag key on which to filter.

tag_key :: Lens' Tag Text

The type of tag on which to filter.

tag_value :: Lens' Tag Text

A value for a tag key on which to filter.

TagFilter

data TagFilter

The tag filter. Valid names are: tagKey, tagValue, configurationId.

See: newTagFilter smart constructor.

Constructors

TagFilter'

Fields

name :: Text

A name of the tag filter.

values :: [Text]

Values for the tag filter.

newTagFilter

Arguments

:: Text
$sel:name:TagFilter'

-> TagFilter

Create a value of TagFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:TagFilter', tagFilter_name - A name of the tag filter.

$sel:values:TagFilter', tagFilter_values - Values for the tag filter.

tagFilter_name :: Lens' TagFilter Text

A name of the tag filter.

tagFilter_values :: Lens' TagFilter [Text]

Values for the tag filter.