Network.AWS.Discovery.Types.AgentInfo

Description

Documentation

data AgentInfo

Information about agents or connectors associated with the user’s AWS account. Information includes agent or connector IDs, IP addresses, media access control (MAC) addresses, agent or connector health, hostname where the agent or connector resides, and agent version for each agent.

See: newAgentInfo smart constructor.

Constructors

AgentInfo'

Fields

hostName :: Maybe Text

The name of the host where the agent or connector resides. The host can be a server or virtual machine.

agentId :: Maybe Text

The agent or connector ID.

agentType :: Maybe Text

Type of agent.

connectorId :: Maybe Text

The ID of the connector.

lastHealthPingTime :: Maybe Text

Time since agent or connector health was reported.

agentNetworkInfoList :: Maybe [AgentNetworkInfo]

Network details about the host where the agent or connector resides.

registeredTime :: Maybe Text

Agent's first registration timestamp in UTC.

version :: Maybe Text

The agent or connector version.

health :: Maybe AgentStatus

The health of the agent or connector.

collectionStatus :: Maybe Text

Status of the collection process for an agent or connector.

newAgentInfo :: AgentInfo

Create a value of AgentInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hostName:AgentInfo', agentInfo_hostName - The name of the host where the agent or connector resides. The host can be a server or virtual machine.

$sel:agentId:AgentInfo', agentInfo_agentId - The agent or connector ID.

$sel:agentType:AgentInfo', agentInfo_agentType - Type of agent.

$sel:connectorId:AgentInfo', agentInfo_connectorId - The ID of the connector.

$sel:lastHealthPingTime:AgentInfo', agentInfo_lastHealthPingTime - Time since agent or connector health was reported.

$sel:agentNetworkInfoList:AgentInfo', agentInfo_agentNetworkInfoList - Network details about the host where the agent or connector resides.

$sel:registeredTime:AgentInfo', agentInfo_registeredTime - Agent's first registration timestamp in UTC.

$sel:version:AgentInfo', agentInfo_version - The agent or connector version.

$sel:health:AgentInfo', agentInfo_health - The health of the agent or connector.

$sel:collectionStatus:AgentInfo', agentInfo_collectionStatus - Status of the collection process for an agent or connector.

agentInfo_hostName :: Lens' AgentInfo (Maybe Text)

The name of the host where the agent or connector resides. The host can be a server or virtual machine.

agentInfo_agentId :: Lens' AgentInfo (Maybe Text)

The agent or connector ID.

agentInfo_agentType :: Lens' AgentInfo (Maybe Text)

Type of agent.

agentInfo_connectorId :: Lens' AgentInfo (Maybe Text)

The ID of the connector.

agentInfo_lastHealthPingTime :: Lens' AgentInfo (Maybe Text)

Time since agent or connector health was reported.

agentInfo_agentNetworkInfoList :: Lens' AgentInfo (Maybe [AgentNetworkInfo])

Network details about the host where the agent or connector resides.

agentInfo_registeredTime :: Lens' AgentInfo (Maybe Text)

Agent's first registration timestamp in UTC.

agentInfo_version :: Lens' AgentInfo (Maybe Text)

The agent or connector version.

agentInfo_health :: Lens' AgentInfo (Maybe AgentStatus)

The health of the agent or connector.

agentInfo_collectionStatus :: Lens' AgentInfo (Maybe Text)

Status of the collection process for an agent or connector.