Network.AWS.Discovery.Types.ConfigurationTag

Description

Documentation

data ConfigurationTag

Tags for a configuration item. Tags are metadata that help you categorize IT assets.

See: newConfigurationTag smart constructor.

Constructors

ConfigurationTag'

Fields

key :: Maybe Text

A type of tag on which to filter. For example, serverType.

configurationId :: Maybe Text

The configuration ID for the item to tag. You can specify a list of keys and values.

value :: Maybe Text

A value on which to filter. For example key = serverType and /value = web server/.

configurationType :: Maybe ConfigurationItemType

A type of IT asset to tag.

timeOfCreation :: Maybe POSIX

The time the configuration tag was created in Coordinated Universal Time (UTC).

newConfigurationTag :: ConfigurationTag

Create a value of ConfigurationTag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:ConfigurationTag', configurationTag_key - A type of tag on which to filter. For example, serverType.

$sel:configurationId:ConfigurationTag', configurationTag_configurationId - The configuration ID for the item to tag. You can specify a list of keys and values.

$sel:value:ConfigurationTag', configurationTag_value - A value on which to filter. For example key = serverType and /value = web server/.

$sel:configurationType:ConfigurationTag', configurationTag_configurationType - A type of IT asset to tag.

$sel:timeOfCreation:ConfigurationTag', configurationTag_timeOfCreation - The time the configuration tag was created in Coordinated Universal Time (UTC).

configurationTag_key :: Lens' ConfigurationTag (Maybe Text)

A type of tag on which to filter. For example, serverType.

configurationTag_configurationId :: Lens' ConfigurationTag (Maybe Text)

The configuration ID for the item to tag. You can specify a list of keys and values.

configurationTag_value :: Lens' ConfigurationTag (Maybe Text)

A value on which to filter. For example key = serverType and /value = web server/.

configurationTag_configurationType :: Lens' ConfigurationTag (Maybe ConfigurationItemType)

A type of IT asset to tag.

configurationTag_timeOfCreation :: Lens' ConfigurationTag (Maybe UTCTime)

The time the configuration tag was created in Coordinated Universal Time (UTC).