Network.AWS.Discovery.Types.ContinuousExportDescription

Description

Documentation

data ContinuousExportDescription

A list of continuous export descriptions.

See: newContinuousExportDescription smart constructor.

Constructors

ContinuousExportDescription'

Fields

status :: Maybe ContinuousExportStatus

Describes the status of the export. Can be one of the following values:

START_IN_PROGRESS - setting up resources to start continuous export.

START_FAILED - an error occurred setting up continuous export. To recover, call start-continuous-export again.

ACTIVE - data is being exported to the customer bucket.

ERROR - an error occurred during export. To fix the issue, call stop-continuous-export and start-continuous-export.

STOP_IN_PROGRESS - stopping the export.

STOP_FAILED - an error occurred stopping the export. To recover, call stop-continuous-export again.

INACTIVE - the continuous export has been stopped. Data is no longer being exported to the customer bucket.

s3Bucket :: Maybe Text

The name of the s3 bucket where the export data parquet files are stored.

stopTime :: Maybe POSIX

The timestamp that represents when this continuous export was stopped.

startTime :: Maybe POSIX

The timestamp representing when the continuous export was started.

dataSource :: Maybe DataSource

The type of data collector used to gather this data (currently only offered for AGENT).

statusDetail :: Maybe Text

Contains information about any errors that have occurred. This data type can have the following values:

ACCESS_DENIED - You don’t have permission to start Data Exploration in Amazon Athena. Contact your AWS administrator for help. For more information, see Setting Up AWS Application Discovery Service in the Application Discovery Service User Guide.

DELIVERY_STREAM_LIMIT_FAILURE - You reached the limit for Amazon Kinesis Data Firehose delivery streams. Reduce the number of streams or request a limit increase and try again. For more information, see Kinesis Data Streams Limits in the Amazon Kinesis Data Streams Developer Guide.

FIREHOSE_ROLE_MISSING - The Data Exploration feature is in an error state because your IAM User is missing the AWSApplicationDiscoveryServiceFirehose role. Turn on Data Exploration in Amazon Athena and try again. For more information, see Step 3: Provide Application Discovery Service Access to Non-Administrator Users by Attaching Policies in the Application Discovery Service User Guide.

FIREHOSE_STREAM_DOES_NOT_EXIST - The Data Exploration feature is in an error state because your IAM User is missing one or more of the Kinesis data delivery streams.

INTERNAL_FAILURE - The Data Exploration feature is in an error state because of an internal failure. Try again later. If this problem persists, contact AWS Support.

S3_BUCKET_LIMIT_FAILURE - You reached the limit for Amazon S3 buckets. Reduce the number of Amazon S3 buckets or request a limit increase and try again. For more information, see Bucket Restrictions and Limitations in the Amazon Simple Storage Service Developer Guide.

S3_NOT_SIGNED_UP - Your account is not signed up for the Amazon S3 service. You must sign up before you can use Amazon S3. You can sign up at the following URL: https://aws.amazon.com/s3.

schemaStorageConfig :: Maybe (HashMap Text Text)

An object which describes how the data is stored.

databaseName - the name of the Glue database used to store the schema.

exportId :: Maybe Text

The unique ID assigned to this export.

newContinuousExportDescription :: ContinuousExportDescription

Create a value of ContinuousExportDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ContinuousExportDescription', continuousExportDescription_status - Describes the status of the export. Can be one of the following values:

START_IN_PROGRESS - setting up resources to start continuous export.

START_FAILED - an error occurred setting up continuous export. To recover, call start-continuous-export again.

ACTIVE - data is being exported to the customer bucket.

ERROR - an error occurred during export. To fix the issue, call stop-continuous-export and start-continuous-export.

STOP_IN_PROGRESS - stopping the export.

STOP_FAILED - an error occurred stopping the export. To recover, call stop-continuous-export again.

INACTIVE - the continuous export has been stopped. Data is no longer being exported to the customer bucket.

$sel:s3Bucket:ContinuousExportDescription', continuousExportDescription_s3Bucket - The name of the s3 bucket where the export data parquet files are stored.

$sel:stopTime:ContinuousExportDescription', continuousExportDescription_stopTime - The timestamp that represents when this continuous export was stopped.

$sel:startTime:ContinuousExportDescription', continuousExportDescription_startTime - The timestamp representing when the continuous export was started.

$sel:dataSource:ContinuousExportDescription', continuousExportDescription_dataSource - The type of data collector used to gather this data (currently only offered for AGENT).

$sel:statusDetail:ContinuousExportDescription', continuousExportDescription_statusDetail - Contains information about any errors that have occurred. This data type can have the following values:

ACCESS_DENIED - You don’t have permission to start Data Exploration in Amazon Athena. Contact your AWS administrator for help. For more information, see Setting Up AWS Application Discovery Service in the Application Discovery Service User Guide.

DELIVERY_STREAM_LIMIT_FAILURE - You reached the limit for Amazon Kinesis Data Firehose delivery streams. Reduce the number of streams or request a limit increase and try again. For more information, see Kinesis Data Streams Limits in the Amazon Kinesis Data Streams Developer Guide.

FIREHOSE_ROLE_MISSING - The Data Exploration feature is in an error state because your IAM User is missing the AWSApplicationDiscoveryServiceFirehose role. Turn on Data Exploration in Amazon Athena and try again. For more information, see Step 3: Provide Application Discovery Service Access to Non-Administrator Users by Attaching Policies in the Application Discovery Service User Guide.

FIREHOSE_STREAM_DOES_NOT_EXIST - The Data Exploration feature is in an error state because your IAM User is missing one or more of the Kinesis data delivery streams.

INTERNAL_FAILURE - The Data Exploration feature is in an error state because of an internal failure. Try again later. If this problem persists, contact AWS Support.

S3_BUCKET_LIMIT_FAILURE - You reached the limit for Amazon S3 buckets. Reduce the number of Amazon S3 buckets or request a limit increase and try again. For more information, see Bucket Restrictions and Limitations in the Amazon Simple Storage Service Developer Guide.

S3_NOT_SIGNED_UP - Your account is not signed up for the Amazon S3 service. You must sign up before you can use Amazon S3. You can sign up at the following URL: https://aws.amazon.com/s3.

$sel:schemaStorageConfig:ContinuousExportDescription', continuousExportDescription_schemaStorageConfig - An object which describes how the data is stored.

databaseName - the name of the Glue database used to store the schema.

$sel:exportId:ContinuousExportDescription', continuousExportDescription_exportId - The unique ID assigned to this export.

continuousExportDescription_status :: Lens' ContinuousExportDescription (Maybe ContinuousExportStatus)

Describes the status of the export. Can be one of the following values:

START_IN_PROGRESS - setting up resources to start continuous export.

START_FAILED - an error occurred setting up continuous export. To recover, call start-continuous-export again.

ACTIVE - data is being exported to the customer bucket.

ERROR - an error occurred during export. To fix the issue, call stop-continuous-export and start-continuous-export.

STOP_IN_PROGRESS - stopping the export.

STOP_FAILED - an error occurred stopping the export. To recover, call stop-continuous-export again.

INACTIVE - the continuous export has been stopped. Data is no longer being exported to the customer bucket.

continuousExportDescription_s3Bucket :: Lens' ContinuousExportDescription (Maybe Text)

The name of the s3 bucket where the export data parquet files are stored.

continuousExportDescription_stopTime :: Lens' ContinuousExportDescription (Maybe UTCTime)

The timestamp that represents when this continuous export was stopped.

continuousExportDescription_startTime :: Lens' ContinuousExportDescription (Maybe UTCTime)

The timestamp representing when the continuous export was started.

continuousExportDescription_dataSource :: Lens' ContinuousExportDescription (Maybe DataSource)

The type of data collector used to gather this data (currently only offered for AGENT).

continuousExportDescription_statusDetail :: Lens' ContinuousExportDescription (Maybe Text)

Contains information about any errors that have occurred. This data type can have the following values:

ACCESS_DENIED - You don’t have permission to start Data Exploration in Amazon Athena. Contact your AWS administrator for help. For more information, see Setting Up AWS Application Discovery Service in the Application Discovery Service User Guide.

DELIVERY_STREAM_LIMIT_FAILURE - You reached the limit for Amazon Kinesis Data Firehose delivery streams. Reduce the number of streams or request a limit increase and try again. For more information, see Kinesis Data Streams Limits in the Amazon Kinesis Data Streams Developer Guide.

FIREHOSE_ROLE_MISSING - The Data Exploration feature is in an error state because your IAM User is missing the AWSApplicationDiscoveryServiceFirehose role. Turn on Data Exploration in Amazon Athena and try again. For more information, see Step 3: Provide Application Discovery Service Access to Non-Administrator Users by Attaching Policies in the Application Discovery Service User Guide.

FIREHOSE_STREAM_DOES_NOT_EXIST - The Data Exploration feature is in an error state because your IAM User is missing one or more of the Kinesis data delivery streams.

INTERNAL_FAILURE - The Data Exploration feature is in an error state because of an internal failure. Try again later. If this problem persists, contact AWS Support.

S3_BUCKET_LIMIT_FAILURE - You reached the limit for Amazon S3 buckets. Reduce the number of Amazon S3 buckets or request a limit increase and try again. For more information, see Bucket Restrictions and Limitations in the Amazon Simple Storage Service Developer Guide.

S3_NOT_SIGNED_UP - Your account is not signed up for the Amazon S3 service. You must sign up before you can use Amazon S3. You can sign up at the following URL: https://aws.amazon.com/s3.

continuousExportDescription_schemaStorageConfig :: Lens' ContinuousExportDescription (Maybe (HashMap Text Text))

An object which describes how the data is stored.

databaseName - the name of the Glue database used to store the schema.

continuousExportDescription_exportId :: Lens' ContinuousExportDescription (Maybe Text)

The unique ID assigned to this export.