Network.AWS.Discovery.Types.NeighborConnectionDetail

Description

Documentation

data NeighborConnectionDetail

Details about neighboring servers.

See: newNeighborConnectionDetail smart constructor.

Constructors

NeighborConnectionDetail'

Fields

transportProtocol :: Maybe Text

The network protocol used for the connection.

destinationPort :: Maybe Int

The destination network port for the connection.

sourceServerId :: Text

The ID of the server that opened the network connection.

destinationServerId :: Text

The ID of the server that accepted the network connection.

connectionsCount :: Integer

The number of open network connections with the neighboring server.

newNeighborConnectionDetail

Arguments

:: Text
$sel:sourceServerId:NeighborConnectionDetail'

-> Text
$sel:destinationServerId:NeighborConnectionDetail'

-> Integer
$sel:connectionsCount:NeighborConnectionDetail'

-> NeighborConnectionDetail

Create a value of NeighborConnectionDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:transportProtocol:NeighborConnectionDetail', neighborConnectionDetail_transportProtocol - The network protocol used for the connection.

$sel:destinationPort:NeighborConnectionDetail', neighborConnectionDetail_destinationPort - The destination network port for the connection.

$sel:sourceServerId:NeighborConnectionDetail', neighborConnectionDetail_sourceServerId - The ID of the server that opened the network connection.

$sel:destinationServerId:NeighborConnectionDetail', neighborConnectionDetail_destinationServerId - The ID of the server that accepted the network connection.

$sel:connectionsCount:NeighborConnectionDetail', neighborConnectionDetail_connectionsCount - The number of open network connections with the neighboring server.

neighborConnectionDetail_transportProtocol :: Lens' NeighborConnectionDetail (Maybe Text)

The network protocol used for the connection.

neighborConnectionDetail_destinationPort :: Lens' NeighborConnectionDetail (Maybe Int)

The destination network port for the connection.

neighborConnectionDetail_sourceServerId :: Lens' NeighborConnectionDetail Text

The ID of the server that opened the network connection.

neighborConnectionDetail_destinationServerId :: Lens' NeighborConnectionDetail Text

The ID of the server that accepted the network connection.

neighborConnectionDetail_connectionsCount :: Lens' NeighborConnectionDetail Integer

The number of open network connections with the neighboring server.