Network.AWS.DocumentDB.CreateDBCluster

Description

Creates a new Amazon DocumentDB cluster.

Creating a Request

data CreateDBCluster

Represents the input to CreateDBCluster.

See: newCreateDBCluster smart constructor.

Constructors

CreateDBCluster'

Fields

engineVersion :: Maybe Text

The version number of the database engine to use. The --engine-version will default to the latest major engine version. For production workloads, we recommend explicitly declaring this parameter with the intended major engine version.

deletionProtection :: Maybe Bool

Specifies whether this cluster can be deleted. If DeletionProtection is enabled, the cluster cannot be deleted unless it is modified and DeletionProtection is disabled. DeletionProtection protects clusters from being accidentally deleted.

storageEncrypted :: Maybe Bool

Specifies whether the cluster is encrypted.

masterUserPassword :: Maybe Text

The password for the master database user. This password can contain any printable ASCII character except forward slash (/), double quote ("), or the "at" symbol (@).

Constraints: Must contain from 8 to 100 characters.

globalClusterIdentifier :: Maybe Text

The cluster identifier of the new global cluster.

masterUsername :: Maybe Text

The name of the master user for the cluster.

Constraints:

Must be from 1 to 63 letters or numbers.

The first character must be a letter.

Cannot be a reserved word for the chosen database engine.

dbSubnetGroupName :: Maybe Text

A subnet group to associate with this cluster.

Constraints: Must match the name of an existing DBSubnetGroup. Must not be default.

Example: mySubnetgroup

preSignedUrl :: Maybe Text

Not currently supported.

preferredMaintenanceWindow :: Maybe Text

The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

Format: ddd:hh24:mi-ddd:hh24:mi

The default is a 30-minute window selected at random from an 8-hour block of time for each Region, occurring on a random day of the week.

Valid days: Mon, Tue, Wed, Thu, Fri, Sat, Sun

Constraints: Minimum 30-minute window.

availabilityZones :: Maybe [Text]

A list of Amazon EC2 Availability Zones that instances in the cluster can be created in.

kmsKeyId :: Maybe Text

The KMS key identifier for an encrypted cluster.

The KMS key identifier is the Amazon Resource Name (ARN) for the KMS encryption key. If you are creating a cluster using the same account that owns the KMS encryption key that is used to encrypt the new cluster, you can use the KMS key alias instead of the ARN for the KMS encryption key.

If an encryption key is not specified in KmsKeyId:

If the StorageEncrypted parameter is true, Amazon DocumentDB uses your default encryption key.

KMS creates the default encryption key for your account. Your account has a different default encryption key for each Regions.

preferredBackupWindow :: Maybe Text

The daily time range during which automated backups are created if automated backups are enabled using the BackupRetentionPeriod parameter.

The default is a 30-minute window selected at random from an 8-hour block of time for each Region.

Constraints:

Must be in the format hh24:mi-hh24:mi.

Must be in Universal Coordinated Time (UTC).

Must not conflict with the preferred maintenance window.

Must be at least 30 minutes.

backupRetentionPeriod :: Maybe Int

The number of days for which automated backups are retained. You must specify a minimum value of 1.

Default: 1

Constraints:

Must be a value from 1 to 35.

vpcSecurityGroupIds :: Maybe [Text]

A list of EC2 VPC security groups to associate with this cluster.

dbClusterParameterGroupName :: Maybe Text

The name of the cluster parameter group to associate with this cluster.

tags :: Maybe [Tag]

The tags to be assigned to the cluster.

port :: Maybe Int

The port number on which the instances in the cluster accept connections.

enableCloudwatchLogsExports :: Maybe [Text]

A list of log types that need to be enabled for exporting to Amazon CloudWatch Logs. You can enable audit logs or profiler logs. For more information, see Auditing Amazon DocumentDB Events and Profiling Amazon DocumentDB Operations.

dbClusterIdentifier :: Text

The cluster identifier. This parameter is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens.

The first character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Example: my-cluster

engine :: Text

The name of the database engine to be used for this cluster.

Valid values: docdb

newCreateDBCluster

Arguments

:: Text
$sel:dbClusterIdentifier:CreateDBCluster'

-> Text
$sel:engine:CreateDBCluster'

-> CreateDBCluster

Create a value of CreateDBCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:engineVersion:CreateDBCluster', createDBCluster_engineVersion - The version number of the database engine to use. The --engine-version will default to the latest major engine version. For production workloads, we recommend explicitly declaring this parameter with the intended major engine version.

$sel:deletionProtection:CreateDBCluster', createDBCluster_deletionProtection - Specifies whether this cluster can be deleted. If DeletionProtection is enabled, the cluster cannot be deleted unless it is modified and DeletionProtection is disabled. DeletionProtection protects clusters from being accidentally deleted.

$sel:storageEncrypted:CreateDBCluster', createDBCluster_storageEncrypted - Specifies whether the cluster is encrypted.

$sel:masterUserPassword:CreateDBCluster', createDBCluster_masterUserPassword - The password for the master database user. This password can contain any printable ASCII character except forward slash (/), double quote ("), or the "at" symbol (@).

Constraints: Must contain from 8 to 100 characters.

$sel:globalClusterIdentifier:CreateDBCluster', createDBCluster_globalClusterIdentifier - The cluster identifier of the new global cluster.

$sel:masterUsername:CreateDBCluster', createDBCluster_masterUsername - The name of the master user for the cluster.

Constraints:

Must be from 1 to 63 letters or numbers.

The first character must be a letter.

Cannot be a reserved word for the chosen database engine.

$sel:dbSubnetGroupName:CreateDBCluster', createDBCluster_dbSubnetGroupName - A subnet group to associate with this cluster.

Constraints: Must match the name of an existing DBSubnetGroup. Must not be default.

Example: mySubnetgroup

$sel:preSignedUrl:CreateDBCluster', createDBCluster_preSignedUrl - Not currently supported.

$sel:preferredMaintenanceWindow:CreateDBCluster', createDBCluster_preferredMaintenanceWindow - The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

Format: ddd:hh24:mi-ddd:hh24:mi

The default is a 30-minute window selected at random from an 8-hour block of time for each Region, occurring on a random day of the week.

Valid days: Mon, Tue, Wed, Thu, Fri, Sat, Sun

Constraints: Minimum 30-minute window.

$sel:availabilityZones:CreateDBCluster', createDBCluster_availabilityZones - A list of Amazon EC2 Availability Zones that instances in the cluster can be created in.

$sel:kmsKeyId:CreateDBCluster', createDBCluster_kmsKeyId - The KMS key identifier for an encrypted cluster.

The KMS key identifier is the Amazon Resource Name (ARN) for the KMS encryption key. If you are creating a cluster using the same account that owns the KMS encryption key that is used to encrypt the new cluster, you can use the KMS key alias instead of the ARN for the KMS encryption key.

If an encryption key is not specified in KmsKeyId:

If the StorageEncrypted parameter is true, Amazon DocumentDB uses your default encryption key.

KMS creates the default encryption key for your account. Your account has a different default encryption key for each Regions.

$sel:preferredBackupWindow:CreateDBCluster', createDBCluster_preferredBackupWindow - The daily time range during which automated backups are created if automated backups are enabled using the BackupRetentionPeriod parameter.

The default is a 30-minute window selected at random from an 8-hour block of time for each Region.

Constraints:

Must be in the format hh24:mi-hh24:mi.

Must be in Universal Coordinated Time (UTC).

Must not conflict with the preferred maintenance window.

Must be at least 30 minutes.

$sel:backupRetentionPeriod:CreateDBCluster', createDBCluster_backupRetentionPeriod - The number of days for which automated backups are retained. You must specify a minimum value of 1.

Default: 1

Constraints:

Must be a value from 1 to 35.

$sel:vpcSecurityGroupIds:CreateDBCluster', createDBCluster_vpcSecurityGroupIds - A list of EC2 VPC security groups to associate with this cluster.

$sel:dbClusterParameterGroupName:CreateDBCluster', createDBCluster_dbClusterParameterGroupName - The name of the cluster parameter group to associate with this cluster.

$sel:tags:CreateDBCluster', createDBCluster_tags - The tags to be assigned to the cluster.

$sel:port:CreateDBCluster', createDBCluster_port - The port number on which the instances in the cluster accept connections.

$sel:enableCloudwatchLogsExports:CreateDBCluster', createDBCluster_enableCloudwatchLogsExports - A list of log types that need to be enabled for exporting to Amazon CloudWatch Logs. You can enable audit logs or profiler logs. For more information, see Auditing Amazon DocumentDB Events and Profiling Amazon DocumentDB Operations.

$sel:dbClusterIdentifier:CreateDBCluster', createDBCluster_dbClusterIdentifier - The cluster identifier. This parameter is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens.

The first character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Example: my-cluster

$sel:engine:CreateDBCluster', createDBCluster_engine - The name of the database engine to be used for this cluster.

Valid values: docdb

Request Lenses

createDBCluster_engineVersion :: Lens' CreateDBCluster (Maybe Text)

The version number of the database engine to use. The --engine-version will default to the latest major engine version. For production workloads, we recommend explicitly declaring this parameter with the intended major engine version.

createDBCluster_deletionProtection :: Lens' CreateDBCluster (Maybe Bool)

Specifies whether this cluster can be deleted. If DeletionProtection is enabled, the cluster cannot be deleted unless it is modified and DeletionProtection is disabled. DeletionProtection protects clusters from being accidentally deleted.

createDBCluster_storageEncrypted :: Lens' CreateDBCluster (Maybe Bool)

Specifies whether the cluster is encrypted.

createDBCluster_masterUserPassword :: Lens' CreateDBCluster (Maybe Text)

The password for the master database user. This password can contain any printable ASCII character except forward slash (/), double quote ("), or the "at" symbol (@).

Constraints: Must contain from 8 to 100 characters.

createDBCluster_globalClusterIdentifier :: Lens' CreateDBCluster (Maybe Text)

The cluster identifier of the new global cluster.

createDBCluster_masterUsername :: Lens' CreateDBCluster (Maybe Text)

The name of the master user for the cluster.

Constraints:

Must be from 1 to 63 letters or numbers.

The first character must be a letter.

Cannot be a reserved word for the chosen database engine.

createDBCluster_dbSubnetGroupName :: Lens' CreateDBCluster (Maybe Text)

A subnet group to associate with this cluster.

Constraints: Must match the name of an existing DBSubnetGroup. Must not be default.

Example: mySubnetgroup

createDBCluster_preSignedUrl :: Lens' CreateDBCluster (Maybe Text)

Not currently supported.

createDBCluster_preferredMaintenanceWindow :: Lens' CreateDBCluster (Maybe Text)

The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

Format: ddd:hh24:mi-ddd:hh24:mi

The default is a 30-minute window selected at random from an 8-hour block of time for each Region, occurring on a random day of the week.

Valid days: Mon, Tue, Wed, Thu, Fri, Sat, Sun

Constraints: Minimum 30-minute window.

createDBCluster_availabilityZones :: Lens' CreateDBCluster (Maybe [Text])

A list of Amazon EC2 Availability Zones that instances in the cluster can be created in.

createDBCluster_kmsKeyId :: Lens' CreateDBCluster (Maybe Text)

The KMS key identifier for an encrypted cluster.

The KMS key identifier is the Amazon Resource Name (ARN) for the KMS encryption key. If you are creating a cluster using the same account that owns the KMS encryption key that is used to encrypt the new cluster, you can use the KMS key alias instead of the ARN for the KMS encryption key.

If an encryption key is not specified in KmsKeyId:

If the StorageEncrypted parameter is true, Amazon DocumentDB uses your default encryption key.

KMS creates the default encryption key for your account. Your account has a different default encryption key for each Regions.

createDBCluster_preferredBackupWindow :: Lens' CreateDBCluster (Maybe Text)

The daily time range during which automated backups are created if automated backups are enabled using the BackupRetentionPeriod parameter.

The default is a 30-minute window selected at random from an 8-hour block of time for each Region.

Constraints:

Must be in the format hh24:mi-hh24:mi.

Must be in Universal Coordinated Time (UTC).

Must not conflict with the preferred maintenance window.

Must be at least 30 minutes.

createDBCluster_backupRetentionPeriod :: Lens' CreateDBCluster (Maybe Int)

The number of days for which automated backups are retained. You must specify a minimum value of 1.

Default: 1

Constraints:

Must be a value from 1 to 35.

createDBCluster_vpcSecurityGroupIds :: Lens' CreateDBCluster (Maybe [Text])

A list of EC2 VPC security groups to associate with this cluster.

createDBCluster_dbClusterParameterGroupName :: Lens' CreateDBCluster (Maybe Text)

The name of the cluster parameter group to associate with this cluster.

createDBCluster_tags :: Lens' CreateDBCluster (Maybe [Tag])

The tags to be assigned to the cluster.

createDBCluster_port :: Lens' CreateDBCluster (Maybe Int)

The port number on which the instances in the cluster accept connections.

createDBCluster_enableCloudwatchLogsExports :: Lens' CreateDBCluster (Maybe [Text])

A list of log types that need to be enabled for exporting to Amazon CloudWatch Logs. You can enable audit logs or profiler logs. For more information, see Auditing Amazon DocumentDB Events and Profiling Amazon DocumentDB Operations.

createDBCluster_dbClusterIdentifier :: Lens' CreateDBCluster Text

The cluster identifier. This parameter is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens.

The first character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Example: my-cluster

createDBCluster_engine :: Lens' CreateDBCluster Text

The name of the database engine to be used for this cluster.

Valid values: docdb

Destructuring the Response

data CreateDBClusterResponse

See: newCreateDBClusterResponse smart constructor.

Constructors

CreateDBClusterResponse'

Fields

dbCluster :: Maybe DBCluster

httpStatus :: Int

The response's http status code.

newCreateDBClusterResponse

Arguments

:: Int
$sel:httpStatus:CreateDBClusterResponse'

-> CreateDBClusterResponse

Create a value of CreateDBClusterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbCluster:CreateDBClusterResponse', createDBClusterResponse_dbCluster - Undocumented member.

$sel:httpStatus:CreateDBClusterResponse', createDBClusterResponse_httpStatus - The response's http status code.

Response Lenses

createDBClusterResponse_dbCluster :: Lens' CreateDBClusterResponse (Maybe DBCluster)

Undocumented member.

createDBClusterResponse_httpStatus :: Lens' CreateDBClusterResponse Int

The response's http status code.