Network.AWS.DocumentDB.DescribeCertificates

Description

Returns a list of certificate authority (CA) certificates provided by Amazon DocumentDB for this account.

This operation returns paginated results.

Creating a Request

data DescribeCertificates

See: newDescribeCertificates smart constructor.

Constructors

DescribeCertificates'

Fields

filters :: Maybe [Filter]

This parameter is not currently supported.

certificateIdentifier :: Maybe Text

The user-supplied certificate identifier. If this parameter is specified, information for only the specified certificate is returned. If this parameter is omitted, a list of up to MaxRecords certificates is returned. This parameter is not case sensitive.

Constraints

Must match an existing CertificateIdentifier.

marker :: Maybe Text

An optional pagination token provided by a previous DescribeCertificates request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

maxRecords :: Maybe Int

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints:

Minimum: 20

Maximum: 100

newDescribeCertificates :: DescribeCertificates

Create a value of DescribeCertificates with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:DescribeCertificates', describeCertificates_filters - This parameter is not currently supported.

$sel:certificateIdentifier:DescribeCertificates', describeCertificates_certificateIdentifier - The user-supplied certificate identifier. If this parameter is specified, information for only the specified certificate is returned. If this parameter is omitted, a list of up to MaxRecords certificates is returned. This parameter is not case sensitive.

Constraints

Must match an existing CertificateIdentifier.

$sel:marker:DescribeCertificates', describeCertificates_marker - An optional pagination token provided by a previous DescribeCertificates request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

$sel:maxRecords:DescribeCertificates', describeCertificates_maxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints:

Minimum: 20

Maximum: 100

Request Lenses

describeCertificates_filters :: Lens' DescribeCertificates (Maybe [Filter])

This parameter is not currently supported.

describeCertificates_certificateIdentifier :: Lens' DescribeCertificates (Maybe Text)

The user-supplied certificate identifier. If this parameter is specified, information for only the specified certificate is returned. If this parameter is omitted, a list of up to MaxRecords certificates is returned. This parameter is not case sensitive.

Constraints

Must match an existing CertificateIdentifier.

describeCertificates_marker :: Lens' DescribeCertificates (Maybe Text)

An optional pagination token provided by a previous DescribeCertificates request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

describeCertificates_maxRecords :: Lens' DescribeCertificates (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints:

Minimum: 20

Maximum: 100

Destructuring the Response

data DescribeCertificatesResponse

See: newDescribeCertificatesResponse smart constructor.

Constructors

DescribeCertificatesResponse'

Fields

certificates :: Maybe [Certificate]

A list of certificates for this account.

marker :: Maybe Text

An optional pagination token provided if the number of records retrieved is greater than MaxRecords. If this parameter is specified, the marker specifies the next record in the list. Including the value of Marker in the next call to DescribeCertificates results in the next page of certificates.

httpStatus :: Int

The response's http status code.

newDescribeCertificatesResponse

Arguments

:: Int
$sel:httpStatus:DescribeCertificatesResponse'

-> DescribeCertificatesResponse

Create a value of DescribeCertificatesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:certificates:DescribeCertificatesResponse', describeCertificatesResponse_certificates - A list of certificates for this account.

$sel:marker:DescribeCertificates', describeCertificatesResponse_marker - An optional pagination token provided if the number of records retrieved is greater than MaxRecords. If this parameter is specified, the marker specifies the next record in the list. Including the value of Marker in the next call to DescribeCertificates results in the next page of certificates.

$sel:httpStatus:DescribeCertificatesResponse', describeCertificatesResponse_httpStatus - The response's http status code.

Response Lenses

describeCertificatesResponse_certificates :: Lens' DescribeCertificatesResponse (Maybe [Certificate])

A list of certificates for this account.

describeCertificatesResponse_marker :: Lens' DescribeCertificatesResponse (Maybe Text)

An optional pagination token provided if the number of records retrieved is greater than MaxRecords. If this parameter is specified, the marker specifies the next record in the list. Including the value of Marker in the next call to DescribeCertificates results in the next page of certificates.

describeCertificatesResponse_httpStatus :: Lens' DescribeCertificatesResponse Int

The response's http status code.