Network.AWS.DocumentDB.DescribeEngineDefaultClusterParameters

Description

Returns the default engine and system parameter information for the cluster database engine.

Creating a Request

data DescribeEngineDefaultClusterParameters

Represents the input to DescribeEngineDefaultClusterParameters.

See: newDescribeEngineDefaultClusterParameters smart constructor.

Constructors

DescribeEngineDefaultClusterParameters'

Fields

filters :: Maybe [Filter]

This parameter is not currently supported.

marker :: Maybe Text

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

maxRecords :: Maybe Int

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token (marker) is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.

dbParameterGroupFamily :: Text

The name of the cluster parameter group family to return the engine parameter information for.

newDescribeEngineDefaultClusterParameters

Arguments

:: Text
$sel:dbParameterGroupFamily:DescribeEngineDefaultClusterParameters'

-> DescribeEngineDefaultClusterParameters

Create a value of DescribeEngineDefaultClusterParameters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:DescribeEngineDefaultClusterParameters', describeEngineDefaultClusterParameters_filters - This parameter is not currently supported.

$sel:marker:DescribeEngineDefaultClusterParameters', describeEngineDefaultClusterParameters_marker - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

$sel:maxRecords:DescribeEngineDefaultClusterParameters', describeEngineDefaultClusterParameters_maxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token (marker) is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.

$sel:dbParameterGroupFamily:DescribeEngineDefaultClusterParameters', describeEngineDefaultClusterParameters_dbParameterGroupFamily - The name of the cluster parameter group family to return the engine parameter information for.

Request Lenses

describeEngineDefaultClusterParameters_filters :: Lens' DescribeEngineDefaultClusterParameters (Maybe [Filter])

This parameter is not currently supported.

describeEngineDefaultClusterParameters_marker :: Lens' DescribeEngineDefaultClusterParameters (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

describeEngineDefaultClusterParameters_maxRecords :: Lens' DescribeEngineDefaultClusterParameters (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token (marker) is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.

describeEngineDefaultClusterParameters_dbParameterGroupFamily :: Lens' DescribeEngineDefaultClusterParameters Text

The name of the cluster parameter group family to return the engine parameter information for.

Destructuring the Response

data DescribeEngineDefaultClusterParametersResponse

See: newDescribeEngineDefaultClusterParametersResponse smart constructor.

Constructors

DescribeEngineDefaultClusterParametersResponse'

Fields

engineDefaults :: Maybe EngineDefaults

httpStatus :: Int

The response's http status code.

newDescribeEngineDefaultClusterParametersResponse

Arguments

:: Int
$sel:httpStatus:DescribeEngineDefaultClusterParametersResponse'

-> DescribeEngineDefaultClusterParametersResponse

Create a value of DescribeEngineDefaultClusterParametersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:engineDefaults:DescribeEngineDefaultClusterParametersResponse', describeEngineDefaultClusterParametersResponse_engineDefaults - Undocumented member.

$sel:httpStatus:DescribeEngineDefaultClusterParametersResponse', describeEngineDefaultClusterParametersResponse_httpStatus - The response's http status code.

Response Lenses

describeEngineDefaultClusterParametersResponse_engineDefaults :: Lens' DescribeEngineDefaultClusterParametersResponse (Maybe EngineDefaults)

Undocumented member.

describeEngineDefaultClusterParametersResponse_httpStatus :: Lens' DescribeEngineDefaultClusterParametersResponse Int

The response's http status code.