Network.AWS.DocumentDB.ModifyDBClusterSnapshotAttribute

Description

Adds an attribute and values to, or removes an attribute and values from, a manual cluster snapshot.

To share a manual cluster snapshot with other accounts, specify restore as the AttributeName, and use the ValuesToAdd parameter to add a list of IDs of the accounts that are authorized to restore the manual cluster snapshot. Use the value all to make the manual cluster snapshot public, which means that it can be copied or restored by all accounts. Do not add the all value for any manual cluster snapshots that contain private information that you don't want available to all accounts. If a manual cluster snapshot is encrypted, it can be shared, but only by specifying a list of authorized account IDs for the ValuesToAdd parameter. You can't use all as a value for that parameter in this case.

Creating a Request

data ModifyDBClusterSnapshotAttribute

Represents the input to ModifyDBClusterSnapshotAttribute.

See: newModifyDBClusterSnapshotAttribute smart constructor.

Constructors

ModifyDBClusterSnapshotAttribute'

Fields

valuesToAdd :: Maybe [Text]

A list of cluster snapshot attributes to add to the attribute specified by AttributeName.

To authorize other accounts to copy or restore a manual cluster snapshot, set this list to include one or more account IDs. To make the manual cluster snapshot restorable by any account, set it to all. Do not add the all value for any manual cluster snapshots that contain private information that you don't want to be available to all accounts.

valuesToRemove :: Maybe [Text]

A list of cluster snapshot attributes to remove from the attribute specified by AttributeName.

To remove authorization for other accounts to copy or restore a manual cluster snapshot, set this list to include one or more account identifiers. To remove authorization for any account to copy or restore the cluster snapshot, set it to all . If you specify all, an account whose account ID is explicitly added to the restore attribute can still copy or restore a manual cluster snapshot.

dbClusterSnapshotIdentifier :: Text

The identifier for the cluster snapshot to modify the attributes for.

attributeName :: Text

The name of the cluster snapshot attribute to modify.

To manage authorization for other accounts to copy or restore a manual cluster snapshot, set this value to restore.

newModifyDBClusterSnapshotAttribute

Arguments

:: Text
$sel:dbClusterSnapshotIdentifier:ModifyDBClusterSnapshotAttribute'

-> Text
$sel:attributeName:ModifyDBClusterSnapshotAttribute'

-> ModifyDBClusterSnapshotAttribute

Create a value of ModifyDBClusterSnapshotAttribute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:valuesToAdd:ModifyDBClusterSnapshotAttribute', modifyDBClusterSnapshotAttribute_valuesToAdd - A list of cluster snapshot attributes to add to the attribute specified by AttributeName.

To authorize other accounts to copy or restore a manual cluster snapshot, set this list to include one or more account IDs. To make the manual cluster snapshot restorable by any account, set it to all. Do not add the all value for any manual cluster snapshots that contain private information that you don't want to be available to all accounts.

$sel:valuesToRemove:ModifyDBClusterSnapshotAttribute', modifyDBClusterSnapshotAttribute_valuesToRemove - A list of cluster snapshot attributes to remove from the attribute specified by AttributeName.

To remove authorization for other accounts to copy or restore a manual cluster snapshot, set this list to include one or more account identifiers. To remove authorization for any account to copy or restore the cluster snapshot, set it to all . If you specify all, an account whose account ID is explicitly added to the restore attribute can still copy or restore a manual cluster snapshot.

$sel:dbClusterSnapshotIdentifier:ModifyDBClusterSnapshotAttribute', modifyDBClusterSnapshotAttribute_dbClusterSnapshotIdentifier - The identifier for the cluster snapshot to modify the attributes for.

$sel:attributeName:ModifyDBClusterSnapshotAttribute', modifyDBClusterSnapshotAttribute_attributeName - The name of the cluster snapshot attribute to modify.

To manage authorization for other accounts to copy or restore a manual cluster snapshot, set this value to restore.

Request Lenses

modifyDBClusterSnapshotAttribute_valuesToAdd :: Lens' ModifyDBClusterSnapshotAttribute (Maybe [Text])

A list of cluster snapshot attributes to add to the attribute specified by AttributeName.

To authorize other accounts to copy or restore a manual cluster snapshot, set this list to include one or more account IDs. To make the manual cluster snapshot restorable by any account, set it to all. Do not add the all value for any manual cluster snapshots that contain private information that you don't want to be available to all accounts.

modifyDBClusterSnapshotAttribute_valuesToRemove :: Lens' ModifyDBClusterSnapshotAttribute (Maybe [Text])

A list of cluster snapshot attributes to remove from the attribute specified by AttributeName.

To remove authorization for other accounts to copy or restore a manual cluster snapshot, set this list to include one or more account identifiers. To remove authorization for any account to copy or restore the cluster snapshot, set it to all . If you specify all, an account whose account ID is explicitly added to the restore attribute can still copy or restore a manual cluster snapshot.

modifyDBClusterSnapshotAttribute_dbClusterSnapshotIdentifier :: Lens' ModifyDBClusterSnapshotAttribute Text

The identifier for the cluster snapshot to modify the attributes for.

modifyDBClusterSnapshotAttribute_attributeName :: Lens' ModifyDBClusterSnapshotAttribute Text

The name of the cluster snapshot attribute to modify.

To manage authorization for other accounts to copy or restore a manual cluster snapshot, set this value to restore.

Destructuring the Response

data ModifyDBClusterSnapshotAttributeResponse

See: newModifyDBClusterSnapshotAttributeResponse smart constructor.

Constructors

ModifyDBClusterSnapshotAttributeResponse'

Fields

dbClusterSnapshotAttributesResult :: Maybe DBClusterSnapshotAttributesResult

httpStatus :: Int

The response's http status code.

newModifyDBClusterSnapshotAttributeResponse

Arguments

:: Int
$sel:httpStatus:ModifyDBClusterSnapshotAttributeResponse'

-> ModifyDBClusterSnapshotAttributeResponse

Create a value of ModifyDBClusterSnapshotAttributeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbClusterSnapshotAttributesResult:ModifyDBClusterSnapshotAttributeResponse', modifyDBClusterSnapshotAttributeResponse_dbClusterSnapshotAttributesResult - Undocumented member.

$sel:httpStatus:ModifyDBClusterSnapshotAttributeResponse', modifyDBClusterSnapshotAttributeResponse_httpStatus - The response's http status code.

Response Lenses

modifyDBClusterSnapshotAttributeResponse_dbClusterSnapshotAttributesResult :: Lens' ModifyDBClusterSnapshotAttributeResponse (Maybe DBClusterSnapshotAttributesResult)

Undocumented member.

modifyDBClusterSnapshotAttributeResponse_httpStatus :: Lens' ModifyDBClusterSnapshotAttributeResponse Int

The response's http status code.