Network.AWS.DocumentDB.ModifyDBInstance

Description

Modifies settings for an instance. You can change one or more database configuration parameters by specifying these parameters and the new values in the request.

Creating a Request

data ModifyDBInstance

Represents the input to ModifyDBInstance.

See: newModifyDBInstance smart constructor.

Constructors

ModifyDBInstance'

Fields

autoMinorVersionUpgrade :: Maybe Bool

This parameter does not apply to Amazon DocumentDB. Amazon DocumentDB does not perform minor version upgrades regardless of the value set.

newDBInstanceIdentifier' :: Maybe Text

The new instance identifier for the instance when renaming an instance. When you change the instance identifier, an instance reboot occurs immediately if you set Apply Immediately to true. It occurs during the next maintenance window if you set Apply Immediately to false. This value is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens.

The first character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Example: mydbinstance

dbInstanceClass :: Maybe Text

The new compute and memory capacity of the instance; for example, db.r5.large. Not all instance classes are available in all Regions.

If you modify the instance class, an outage occurs during the change. The change is applied during the next maintenance window, unless ApplyImmediately is specified as true for this request.

Default: Uses existing setting.

promotionTier :: Maybe Int

A value that specifies the order in which an Amazon DocumentDB replica is promoted to the primary instance after a failure of the existing primary instance.

Default: 1

Valid values: 0-15

preferredMaintenanceWindow :: Maybe Text

The weekly time range (in UTC) during which system maintenance can occur, which might result in an outage. Changing this parameter doesn't result in an outage except in the following situation, and the change is asynchronously applied as soon as possible. If there are pending actions that cause a reboot, and the maintenance window is changed to include the current time, changing this parameter causes a reboot of the instance. If you are moving this window to the current time, there must be at least 30 minutes between the current time and end of the window to ensure that pending changes are applied.

Default: Uses existing setting.

Format: ddd:hh24:mi-ddd:hh24:mi

Valid days: Mon, Tue, Wed, Thu, Fri, Sat, Sun

Constraints: Must be at least 30 minutes.

cACertificateIdentifier :: Maybe Text

Indicates the certificate that needs to be associated with the instance.

applyImmediately :: Maybe Bool

Specifies whether the modifications in this request and any pending modifications are asynchronously applied as soon as possible, regardless of the PreferredMaintenanceWindow setting for the instance.

If this parameter is set to false, changes to the instance are applied during the next maintenance window. Some parameter changes can cause an outage and are applied on the next reboot.

Default: false

dbInstanceIdentifier :: Text

The instance identifier. This value is stored as a lowercase string.

Constraints:

Must match the identifier of an existing DBInstance.

newModifyDBInstance

Arguments

:: Text
$sel:dbInstanceIdentifier:ModifyDBInstance'

-> ModifyDBInstance

Create a value of ModifyDBInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:autoMinorVersionUpgrade:ModifyDBInstance', modifyDBInstance_autoMinorVersionUpgrade - This parameter does not apply to Amazon DocumentDB. Amazon DocumentDB does not perform minor version upgrades regardless of the value set.

$sel:newDBInstanceIdentifier':ModifyDBInstance', modifyDBInstance_newDBInstanceIdentifier - The new instance identifier for the instance when renaming an instance. When you change the instance identifier, an instance reboot occurs immediately if you set Apply Immediately to true. It occurs during the next maintenance window if you set Apply Immediately to false. This value is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens.

The first character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Example: mydbinstance

$sel:dbInstanceClass:ModifyDBInstance', modifyDBInstance_dbInstanceClass - The new compute and memory capacity of the instance; for example, db.r5.large. Not all instance classes are available in all Regions.

If you modify the instance class, an outage occurs during the change. The change is applied during the next maintenance window, unless ApplyImmediately is specified as true for this request.

Default: Uses existing setting.

$sel:promotionTier:ModifyDBInstance', modifyDBInstance_promotionTier - A value that specifies the order in which an Amazon DocumentDB replica is promoted to the primary instance after a failure of the existing primary instance.

Default: 1

Valid values: 0-15

$sel:preferredMaintenanceWindow:ModifyDBInstance', modifyDBInstance_preferredMaintenanceWindow - The weekly time range (in UTC) during which system maintenance can occur, which might result in an outage. Changing this parameter doesn't result in an outage except in the following situation, and the change is asynchronously applied as soon as possible. If there are pending actions that cause a reboot, and the maintenance window is changed to include the current time, changing this parameter causes a reboot of the instance. If you are moving this window to the current time, there must be at least 30 minutes between the current time and end of the window to ensure that pending changes are applied.

Default: Uses existing setting.

Format: ddd:hh24:mi-ddd:hh24:mi

Valid days: Mon, Tue, Wed, Thu, Fri, Sat, Sun

Constraints: Must be at least 30 minutes.

$sel:cACertificateIdentifier:ModifyDBInstance', modifyDBInstance_cACertificateIdentifier - Indicates the certificate that needs to be associated with the instance.

$sel:applyImmediately:ModifyDBInstance', modifyDBInstance_applyImmediately - Specifies whether the modifications in this request and any pending modifications are asynchronously applied as soon as possible, regardless of the PreferredMaintenanceWindow setting for the instance.

If this parameter is set to false, changes to the instance are applied during the next maintenance window. Some parameter changes can cause an outage and are applied on the next reboot.

Default: false

$sel:dbInstanceIdentifier:ModifyDBInstance', modifyDBInstance_dbInstanceIdentifier - The instance identifier. This value is stored as a lowercase string.

Constraints:

Must match the identifier of an existing DBInstance.

Request Lenses

modifyDBInstance_autoMinorVersionUpgrade :: Lens' ModifyDBInstance (Maybe Bool)

This parameter does not apply to Amazon DocumentDB. Amazon DocumentDB does not perform minor version upgrades regardless of the value set.

modifyDBInstance_newDBInstanceIdentifier :: Lens' ModifyDBInstance (Maybe Text)

The new instance identifier for the instance when renaming an instance. When you change the instance identifier, an instance reboot occurs immediately if you set Apply Immediately to true. It occurs during the next maintenance window if you set Apply Immediately to false. This value is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens.

The first character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Example: mydbinstance

modifyDBInstance_dbInstanceClass :: Lens' ModifyDBInstance (Maybe Text)

The new compute and memory capacity of the instance; for example, db.r5.large. Not all instance classes are available in all Regions.

If you modify the instance class, an outage occurs during the change. The change is applied during the next maintenance window, unless ApplyImmediately is specified as true for this request.

Default: Uses existing setting.

modifyDBInstance_promotionTier :: Lens' ModifyDBInstance (Maybe Int)

A value that specifies the order in which an Amazon DocumentDB replica is promoted to the primary instance after a failure of the existing primary instance.

Default: 1

Valid values: 0-15

modifyDBInstance_preferredMaintenanceWindow :: Lens' ModifyDBInstance (Maybe Text)

The weekly time range (in UTC) during which system maintenance can occur, which might result in an outage. Changing this parameter doesn't result in an outage except in the following situation, and the change is asynchronously applied as soon as possible. If there are pending actions that cause a reboot, and the maintenance window is changed to include the current time, changing this parameter causes a reboot of the instance. If you are moving this window to the current time, there must be at least 30 minutes between the current time and end of the window to ensure that pending changes are applied.

Default: Uses existing setting.

Format: ddd:hh24:mi-ddd:hh24:mi

Valid days: Mon, Tue, Wed, Thu, Fri, Sat, Sun

Constraints: Must be at least 30 minutes.

modifyDBInstance_cACertificateIdentifier :: Lens' ModifyDBInstance (Maybe Text)

Indicates the certificate that needs to be associated with the instance.

modifyDBInstance_applyImmediately :: Lens' ModifyDBInstance (Maybe Bool)

Specifies whether the modifications in this request and any pending modifications are asynchronously applied as soon as possible, regardless of the PreferredMaintenanceWindow setting for the instance.

If this parameter is set to false, changes to the instance are applied during the next maintenance window. Some parameter changes can cause an outage and are applied on the next reboot.

Default: false

modifyDBInstance_dbInstanceIdentifier :: Lens' ModifyDBInstance Text

The instance identifier. This value is stored as a lowercase string.

Constraints:

Must match the identifier of an existing DBInstance.

Destructuring the Response

data ModifyDBInstanceResponse

See: newModifyDBInstanceResponse smart constructor.

Constructors

ModifyDBInstanceResponse'

Fields

dbInstance :: Maybe DBInstance

httpStatus :: Int

The response's http status code.

newModifyDBInstanceResponse

Arguments

:: Int
$sel:httpStatus:ModifyDBInstanceResponse'

-> ModifyDBInstanceResponse

Create a value of ModifyDBInstanceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbInstance:ModifyDBInstanceResponse', modifyDBInstanceResponse_dbInstance - Undocumented member.

$sel:httpStatus:ModifyDBInstanceResponse', modifyDBInstanceResponse_httpStatus - The response's http status code.

Response Lenses

modifyDBInstanceResponse_dbInstance :: Lens' ModifyDBInstanceResponse (Maybe DBInstance)

Undocumented member.

modifyDBInstanceResponse_httpStatus :: Lens' ModifyDBInstanceResponse Int

The response's http status code.