Network.AWS.DocumentDB.ModifyDBSubnetGroup

Description

Modifies an existing subnet group. subnet groups must contain at least one subnet in at least two Availability Zones in the Region.

Creating a Request

data ModifyDBSubnetGroup

Represents the input to ModifyDBSubnetGroup.

See: newModifyDBSubnetGroup smart constructor.

Constructors

ModifyDBSubnetGroup'

Fields

dbSubnetGroupDescription :: Maybe Text

The description for the subnet group.

dbSubnetGroupName :: Text

The name for the subnet group. This value is stored as a lowercase string. You can't modify the default subnet group.

Constraints: Must match the name of an existing DBSubnetGroup. Must not be default.

Example: mySubnetgroup

subnetIds :: [Text]

The Amazon EC2 subnet IDs for the subnet group.

newModifyDBSubnetGroup

Arguments

:: Text
$sel:dbSubnetGroupName:ModifyDBSubnetGroup'

-> ModifyDBSubnetGroup

Create a value of ModifyDBSubnetGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbSubnetGroupDescription:ModifyDBSubnetGroup', modifyDBSubnetGroup_dbSubnetGroupDescription - The description for the subnet group.

$sel:dbSubnetGroupName:ModifyDBSubnetGroup', modifyDBSubnetGroup_dbSubnetGroupName - The name for the subnet group. This value is stored as a lowercase string. You can't modify the default subnet group.

Constraints: Must match the name of an existing DBSubnetGroup. Must not be default.

Example: mySubnetgroup

$sel:subnetIds:ModifyDBSubnetGroup', modifyDBSubnetGroup_subnetIds - The Amazon EC2 subnet IDs for the subnet group.

Request Lenses

modifyDBSubnetGroup_dbSubnetGroupDescription :: Lens' ModifyDBSubnetGroup (Maybe Text)

The description for the subnet group.

modifyDBSubnetGroup_dbSubnetGroupName :: Lens' ModifyDBSubnetGroup Text

The name for the subnet group. This value is stored as a lowercase string. You can't modify the default subnet group.

Constraints: Must match the name of an existing DBSubnetGroup. Must not be default.

Example: mySubnetgroup

modifyDBSubnetGroup_subnetIds :: Lens' ModifyDBSubnetGroup [Text]

The Amazon EC2 subnet IDs for the subnet group.

Destructuring the Response

data ModifyDBSubnetGroupResponse

See: newModifyDBSubnetGroupResponse smart constructor.

Constructors

ModifyDBSubnetGroupResponse'

Fields

dbSubnetGroup :: Maybe DBSubnetGroup

httpStatus :: Int

The response's http status code.

newModifyDBSubnetGroupResponse

Arguments

:: Int
$sel:httpStatus:ModifyDBSubnetGroupResponse'

-> ModifyDBSubnetGroupResponse

Create a value of ModifyDBSubnetGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbSubnetGroup:ModifyDBSubnetGroupResponse', modifyDBSubnetGroupResponse_dbSubnetGroup - Undocumented member.

$sel:httpStatus:ModifyDBSubnetGroupResponse', modifyDBSubnetGroupResponse_httpStatus - The response's http status code.

Response Lenses

modifyDBSubnetGroupResponse_dbSubnetGroup :: Lens' ModifyDBSubnetGroupResponse (Maybe DBSubnetGroup)

Undocumented member.

modifyDBSubnetGroupResponse_httpStatus :: Lens' ModifyDBSubnetGroupResponse Int

The response's http status code.