Network.AWS.DocumentDB.RebootDBInstance

Description

You might need to reboot your instance, usually for maintenance reasons. For example, if you make certain changes, or if you change the cluster parameter group that is associated with the instance, you must reboot the instance for the changes to take effect.

Rebooting an instance restarts the database engine service. Rebooting an instance results in a momentary outage, during which the instance status is set to rebooting.

Creating a Request

data RebootDBInstance

Represents the input to RebootDBInstance.

See: newRebootDBInstance smart constructor.

Constructors

RebootDBInstance'

Fields

forceFailover :: Maybe Bool

When true, the reboot is conducted through a Multi-AZ failover.

Constraint: You can't specify true if the instance is not configured for Multi-AZ.

dbInstanceIdentifier :: Text

The instance identifier. This parameter is stored as a lowercase string.

Constraints:

Must match the identifier of an existing DBInstance.

newRebootDBInstance

Arguments

:: Text
$sel:dbInstanceIdentifier:RebootDBInstance'

-> RebootDBInstance

Create a value of RebootDBInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:forceFailover:RebootDBInstance', rebootDBInstance_forceFailover - When true, the reboot is conducted through a Multi-AZ failover.

Constraint: You can't specify true if the instance is not configured for Multi-AZ.

$sel:dbInstanceIdentifier:RebootDBInstance', rebootDBInstance_dbInstanceIdentifier - The instance identifier. This parameter is stored as a lowercase string.

Constraints:

Must match the identifier of an existing DBInstance.

Request Lenses

rebootDBInstance_forceFailover :: Lens' RebootDBInstance (Maybe Bool)

When true, the reboot is conducted through a Multi-AZ failover.

Constraint: You can't specify true if the instance is not configured for Multi-AZ.

rebootDBInstance_dbInstanceIdentifier :: Lens' RebootDBInstance Text

The instance identifier. This parameter is stored as a lowercase string.

Constraints:

Must match the identifier of an existing DBInstance.

Destructuring the Response

data RebootDBInstanceResponse

See: newRebootDBInstanceResponse smart constructor.

Constructors

RebootDBInstanceResponse'

Fields

dbInstance :: Maybe DBInstance

httpStatus :: Int

The response's http status code.

newRebootDBInstanceResponse

Arguments

:: Int
$sel:httpStatus:RebootDBInstanceResponse'

-> RebootDBInstanceResponse

Create a value of RebootDBInstanceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbInstance:RebootDBInstanceResponse', rebootDBInstanceResponse_dbInstance - Undocumented member.

$sel:httpStatus:RebootDBInstanceResponse', rebootDBInstanceResponse_httpStatus - The response's http status code.

Response Lenses

rebootDBInstanceResponse_dbInstance :: Lens' RebootDBInstanceResponse (Maybe DBInstance)

Undocumented member.

rebootDBInstanceResponse_httpStatus :: Lens' RebootDBInstanceResponse Int

The response's http status code.