Network.AWS.DocumentDB.RestoreDBClusterToPointInTime

Description

Restores a cluster to an arbitrary point in time. Users can restore to any point in time before LatestRestorableTime for up to BackupRetentionPeriod days. The target cluster is created from the source cluster with the same configuration as the original cluster, except that the new cluster is created with the default security group.

Creating a Request

data RestoreDBClusterToPointInTime

Represents the input to RestoreDBClusterToPointInTime.

See: newRestoreDBClusterToPointInTime smart constructor.

Constructors

RestoreDBClusterToPointInTime'

Fields

deletionProtection :: Maybe Bool

Specifies whether this cluster can be deleted. If DeletionProtection is enabled, the cluster cannot be deleted unless it is modified and DeletionProtection is disabled. DeletionProtection protects clusters from being accidentally deleted.

useLatestRestorableTime :: Maybe Bool

A value that is set to true to restore the cluster to the latest restorable backup time, and false otherwise.

Default: false

Constraints: Cannot be specified if the RestoreToTime parameter is provided.

dbSubnetGroupName :: Maybe Text

The subnet group name to use for the new cluster.

Constraints: If provided, must match the name of an existing DBSubnetGroup.

Example: mySubnetgroup

kmsKeyId :: Maybe Text

The KMS key identifier to use when restoring an encrypted cluster from an encrypted cluster.

The KMS key identifier is the Amazon Resource Name (ARN) for the KMS encryption key. If you are restoring a cluster with the same account that owns the KMS encryption key used to encrypt the new cluster, then you can use the KMS key alias instead of the ARN for the KMS encryption key.

You can restore to a new cluster and encrypt the new cluster with an KMS key that is different from the KMS key used to encrypt the source cluster. The new DB cluster is encrypted with the KMS key identified by the KmsKeyId parameter.

If you do not specify a value for the KmsKeyId parameter, then the following occurs:

If the cluster is encrypted, then the restored cluster is encrypted using the KMS key that was used to encrypt the source cluster.

If the cluster is not encrypted, then the restored cluster is not encrypted.

If DBClusterIdentifier refers to a cluster that is not encrypted, then the restore request is rejected.

vpcSecurityGroupIds :: Maybe [Text]

A list of VPC security groups that the new cluster belongs to.

restoreToTime :: Maybe ISO8601

The date and time to restore the cluster to.

Valid values: A time in Universal Coordinated Time (UTC) format.

Constraints:

Must be before the latest restorable time for the instance.

Must be specified if the UseLatestRestorableTime parameter is not provided.

Cannot be specified if the UseLatestRestorableTime parameter is true.

Cannot be specified if the RestoreType parameter is copy-on-write.

Example: 2015-03-07T23:45:00Z

tags :: Maybe [Tag]

The tags to be assigned to the restored cluster.

port :: Maybe Int

The port number on which the new cluster accepts connections.

Constraints: Must be a value from 1150 to 65535.

Default: The default port for the engine.

enableCloudwatchLogsExports :: Maybe [Text]

A list of log types that must be enabled for exporting to Amazon CloudWatch Logs.

dbClusterIdentifier :: Text

The name of the new cluster to be created.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens.

The first character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

sourceDBClusterIdentifier :: Text

The identifier of the source cluster from which to restore.

Constraints:

Must match the identifier of an existing DBCluster.

newRestoreDBClusterToPointInTime

Arguments

:: Text
$sel:dbClusterIdentifier:RestoreDBClusterToPointInTime'

-> Text
$sel:sourceDBClusterIdentifier:RestoreDBClusterToPointInTime'

-> RestoreDBClusterToPointInTime

Create a value of RestoreDBClusterToPointInTime with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deletionProtection:RestoreDBClusterToPointInTime', restoreDBClusterToPointInTime_deletionProtection - Specifies whether this cluster can be deleted. If DeletionProtection is enabled, the cluster cannot be deleted unless it is modified and DeletionProtection is disabled. DeletionProtection protects clusters from being accidentally deleted.

$sel:useLatestRestorableTime:RestoreDBClusterToPointInTime', restoreDBClusterToPointInTime_useLatestRestorableTime - A value that is set to true to restore the cluster to the latest restorable backup time, and false otherwise.

Default: false

Constraints: Cannot be specified if the RestoreToTime parameter is provided.

$sel:dbSubnetGroupName:RestoreDBClusterToPointInTime', restoreDBClusterToPointInTime_dbSubnetGroupName - The subnet group name to use for the new cluster.

Constraints: If provided, must match the name of an existing DBSubnetGroup.

Example: mySubnetgroup

$sel:kmsKeyId:RestoreDBClusterToPointInTime', restoreDBClusterToPointInTime_kmsKeyId - The KMS key identifier to use when restoring an encrypted cluster from an encrypted cluster.

The KMS key identifier is the Amazon Resource Name (ARN) for the KMS encryption key. If you are restoring a cluster with the same account that owns the KMS encryption key used to encrypt the new cluster, then you can use the KMS key alias instead of the ARN for the KMS encryption key.

You can restore to a new cluster and encrypt the new cluster with an KMS key that is different from the KMS key used to encrypt the source cluster. The new DB cluster is encrypted with the KMS key identified by the KmsKeyId parameter.

If you do not specify a value for the KmsKeyId parameter, then the following occurs:

If the cluster is encrypted, then the restored cluster is encrypted using the KMS key that was used to encrypt the source cluster.

If the cluster is not encrypted, then the restored cluster is not encrypted.

If DBClusterIdentifier refers to a cluster that is not encrypted, then the restore request is rejected.

$sel:vpcSecurityGroupIds:RestoreDBClusterToPointInTime', restoreDBClusterToPointInTime_vpcSecurityGroupIds - A list of VPC security groups that the new cluster belongs to.

$sel:restoreToTime:RestoreDBClusterToPointInTime', restoreDBClusterToPointInTime_restoreToTime - The date and time to restore the cluster to.

Valid values: A time in Universal Coordinated Time (UTC) format.

Constraints:

Must be before the latest restorable time for the instance.

Must be specified if the UseLatestRestorableTime parameter is not provided.

Cannot be specified if the UseLatestRestorableTime parameter is true.

Cannot be specified if the RestoreType parameter is copy-on-write.

Example: 2015-03-07T23:45:00Z

$sel:tags:RestoreDBClusterToPointInTime', restoreDBClusterToPointInTime_tags - The tags to be assigned to the restored cluster.

$sel:port:RestoreDBClusterToPointInTime', restoreDBClusterToPointInTime_port - The port number on which the new cluster accepts connections.

Constraints: Must be a value from 1150 to 65535.

Default: The default port for the engine.

$sel:enableCloudwatchLogsExports:RestoreDBClusterToPointInTime', restoreDBClusterToPointInTime_enableCloudwatchLogsExports - A list of log types that must be enabled for exporting to Amazon CloudWatch Logs.

$sel:dbClusterIdentifier:RestoreDBClusterToPointInTime', restoreDBClusterToPointInTime_dbClusterIdentifier - The name of the new cluster to be created.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens.

The first character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

$sel:sourceDBClusterIdentifier:RestoreDBClusterToPointInTime', restoreDBClusterToPointInTime_sourceDBClusterIdentifier - The identifier of the source cluster from which to restore.

Constraints:

Must match the identifier of an existing DBCluster.

Request Lenses

restoreDBClusterToPointInTime_deletionProtection :: Lens' RestoreDBClusterToPointInTime (Maybe Bool)

Specifies whether this cluster can be deleted. If DeletionProtection is enabled, the cluster cannot be deleted unless it is modified and DeletionProtection is disabled. DeletionProtection protects clusters from being accidentally deleted.

restoreDBClusterToPointInTime_useLatestRestorableTime :: Lens' RestoreDBClusterToPointInTime (Maybe Bool)

A value that is set to true to restore the cluster to the latest restorable backup time, and false otherwise.

Default: false

Constraints: Cannot be specified if the RestoreToTime parameter is provided.

restoreDBClusterToPointInTime_dbSubnetGroupName :: Lens' RestoreDBClusterToPointInTime (Maybe Text)

The subnet group name to use for the new cluster.

Constraints: If provided, must match the name of an existing DBSubnetGroup.

Example: mySubnetgroup

restoreDBClusterToPointInTime_kmsKeyId :: Lens' RestoreDBClusterToPointInTime (Maybe Text)

The KMS key identifier to use when restoring an encrypted cluster from an encrypted cluster.

The KMS key identifier is the Amazon Resource Name (ARN) for the KMS encryption key. If you are restoring a cluster with the same account that owns the KMS encryption key used to encrypt the new cluster, then you can use the KMS key alias instead of the ARN for the KMS encryption key.

You can restore to a new cluster and encrypt the new cluster with an KMS key that is different from the KMS key used to encrypt the source cluster. The new DB cluster is encrypted with the KMS key identified by the KmsKeyId parameter.

If you do not specify a value for the KmsKeyId parameter, then the following occurs:

If the cluster is encrypted, then the restored cluster is encrypted using the KMS key that was used to encrypt the source cluster.

If the cluster is not encrypted, then the restored cluster is not encrypted.

If DBClusterIdentifier refers to a cluster that is not encrypted, then the restore request is rejected.

restoreDBClusterToPointInTime_vpcSecurityGroupIds :: Lens' RestoreDBClusterToPointInTime (Maybe [Text])

A list of VPC security groups that the new cluster belongs to.

restoreDBClusterToPointInTime_restoreToTime :: Lens' RestoreDBClusterToPointInTime (Maybe UTCTime)

The date and time to restore the cluster to.

Valid values: A time in Universal Coordinated Time (UTC) format.

Constraints:

Must be before the latest restorable time for the instance.

Must be specified if the UseLatestRestorableTime parameter is not provided.

Cannot be specified if the UseLatestRestorableTime parameter is true.

Cannot be specified if the RestoreType parameter is copy-on-write.

Example: 2015-03-07T23:45:00Z

restoreDBClusterToPointInTime_tags :: Lens' RestoreDBClusterToPointInTime (Maybe [Tag])

The tags to be assigned to the restored cluster.

restoreDBClusterToPointInTime_port :: Lens' RestoreDBClusterToPointInTime (Maybe Int)

The port number on which the new cluster accepts connections.

Constraints: Must be a value from 1150 to 65535.

Default: The default port for the engine.

restoreDBClusterToPointInTime_enableCloudwatchLogsExports :: Lens' RestoreDBClusterToPointInTime (Maybe [Text])

A list of log types that must be enabled for exporting to Amazon CloudWatch Logs.

restoreDBClusterToPointInTime_dbClusterIdentifier :: Lens' RestoreDBClusterToPointInTime Text

The name of the new cluster to be created.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens.

The first character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

restoreDBClusterToPointInTime_sourceDBClusterIdentifier :: Lens' RestoreDBClusterToPointInTime Text

The identifier of the source cluster from which to restore.

Constraints:

Must match the identifier of an existing DBCluster.

Destructuring the Response

data RestoreDBClusterToPointInTimeResponse

See: newRestoreDBClusterToPointInTimeResponse smart constructor.

Constructors

RestoreDBClusterToPointInTimeResponse'

Fields

dbCluster :: Maybe DBCluster

httpStatus :: Int

The response's http status code.

newRestoreDBClusterToPointInTimeResponse

Arguments

:: Int
$sel:httpStatus:RestoreDBClusterToPointInTimeResponse'

-> RestoreDBClusterToPointInTimeResponse

Create a value of RestoreDBClusterToPointInTimeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbCluster:RestoreDBClusterToPointInTimeResponse', restoreDBClusterToPointInTimeResponse_dbCluster - Undocumented member.

$sel:httpStatus:RestoreDBClusterToPointInTimeResponse', restoreDBClusterToPointInTimeResponse_httpStatus - The response's http status code.

Response Lenses

restoreDBClusterToPointInTimeResponse_dbCluster :: Lens' RestoreDBClusterToPointInTimeResponse (Maybe DBCluster)

Undocumented member.

restoreDBClusterToPointInTimeResponse_httpStatus :: Lens' RestoreDBClusterToPointInTimeResponse Int

The response's http status code.