Network.AWS.DocumentDB.Types.DBCluster

Description

Documentation

data DBCluster

Detailed information about a cluster.

See: newDBCluster smart constructor.

Constructors

DBCluster'

Fields

engineVersion :: Maybe Text

Indicates the database engine version.

status :: Maybe Text

Specifies the current state of this cluster.

deletionProtection :: Maybe Bool

Specifies whether this cluster can be deleted. If DeletionProtection is enabled, the cluster cannot be deleted unless it is modified and DeletionProtection is disabled. DeletionProtection protects clusters from being accidentally deleted.

storageEncrypted :: Maybe Bool

Specifies whether the cluster is encrypted.

dbClusterIdentifier :: Maybe Text

Contains a user-supplied cluster identifier. This identifier is the unique key that identifies a cluster.

dbClusterMembers :: Maybe [DBClusterMember]

Provides the list of instances that make up the cluster.

readReplicaIdentifiers :: Maybe [Text]

Contains one or more identifiers of the secondary clusters that are associated with this cluster.

replicationSourceIdentifier :: Maybe Text

Contains the identifier of the source cluster if this cluster is a secondary cluster.

hostedZoneId :: Maybe Text

Specifies the ID that Amazon Route 53 assigns when you create a hosted zone.

dbClusterParameterGroup :: Maybe Text

Specifies the name of the cluster parameter group for the cluster.

masterUsername :: Maybe Text

Contains the master user name for the cluster.

dbClusterResourceId :: Maybe Text

The Region-unique, immutable identifier for the cluster. This identifier is found in CloudTrail log entries whenever the KMS key for the cluster is accessed.

earliestRestorableTime :: Maybe ISO8601

The earliest time to which a database can be restored with point-in-time restore.

engine :: Maybe Text

Provides the name of the database engine to be used for this cluster.

dbClusterArn :: Maybe Text

The Amazon Resource Name (ARN) for the cluster.

latestRestorableTime :: Maybe ISO8601

Specifies the latest time to which a database can be restored with point-in-time restore.

preferredMaintenanceWindow :: Maybe Text

Specifies the weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

availabilityZones :: Maybe [Text]

Provides the list of Amazon EC2 Availability Zones that instances in the cluster can be created in.

kmsKeyId :: Maybe Text

If StorageEncrypted is true, the KMS key identifier for the encrypted cluster.

preferredBackupWindow :: Maybe Text

Specifies the daily time range during which automated backups are created if automated backups are enabled, as determined by the BackupRetentionPeriod.

associatedRoles :: Maybe [DBClusterRole]

Provides a list of the Identity and Access Management (IAM) roles that are associated with the cluster. (IAM) roles that are associated with a cluster grant permission for the cluster to access other Amazon Web Services services on your behalf.

vpcSecurityGroups :: Maybe [VpcSecurityGroupMembership]

Provides a list of virtual private cloud (VPC) security groups that the cluster belongs to.

backupRetentionPeriod :: Maybe Int

Specifies the number of days for which automatic snapshots are retained.

dbSubnetGroup :: Maybe Text

Specifies information on the subnet group that is associated with the cluster, including the name, description, and subnets in the subnet group.

multiAZ :: Maybe Bool

Specifies whether the cluster has instances in multiple Availability Zones.

enabledCloudwatchLogsExports :: Maybe [Text]

A list of log types that this cluster is configured to export to Amazon CloudWatch Logs.

clusterCreateTime :: Maybe ISO8601

Specifies the time when the cluster was created, in Universal Coordinated Time (UTC).

endpoint :: Maybe Text

Specifies the connection endpoint for the primary instance of the cluster.

percentProgress :: Maybe Text

Specifies the progress of the operation as a percentage.

readerEndpoint :: Maybe Text

The reader endpoint for the cluster. The reader endpoint for a cluster load balances connections across the Amazon DocumentDB replicas that are available in a cluster. As clients request new connections to the reader endpoint, Amazon DocumentDB distributes the connection requests among the Amazon DocumentDB replicas in the cluster. This functionality can help balance your read workload across multiple Amazon DocumentDB replicas in your cluster.

If a failover occurs, and the Amazon DocumentDB replica that you are connected to is promoted to be the primary instance, your connection is dropped. To continue sending your read workload to other Amazon DocumentDB replicas in the cluster, you can then reconnect to the reader endpoint.

port :: Maybe Int

Specifies the port that the database engine is listening on.

newDBCluster :: DBCluster

Create a value of DBCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:engineVersion:DBCluster', dbCluster_engineVersion - Indicates the database engine version.

$sel:status:DBCluster', dbCluster_status - Specifies the current state of this cluster.

$sel:deletionProtection:DBCluster', dbCluster_deletionProtection - Specifies whether this cluster can be deleted. If DeletionProtection is enabled, the cluster cannot be deleted unless it is modified and DeletionProtection is disabled. DeletionProtection protects clusters from being accidentally deleted.

$sel:storageEncrypted:DBCluster', dbCluster_storageEncrypted - Specifies whether the cluster is encrypted.

$sel:dbClusterIdentifier:DBCluster', dbCluster_dbClusterIdentifier - Contains a user-supplied cluster identifier. This identifier is the unique key that identifies a cluster.

$sel:dbClusterMembers:DBCluster', dbCluster_dbClusterMembers - Provides the list of instances that make up the cluster.

$sel:readReplicaIdentifiers:DBCluster', dbCluster_readReplicaIdentifiers - Contains one or more identifiers of the secondary clusters that are associated with this cluster.

$sel:replicationSourceIdentifier:DBCluster', dbCluster_replicationSourceIdentifier - Contains the identifier of the source cluster if this cluster is a secondary cluster.

$sel:hostedZoneId:DBCluster', dbCluster_hostedZoneId - Specifies the ID that Amazon Route 53 assigns when you create a hosted zone.

$sel:dbClusterParameterGroup:DBCluster', dbCluster_dbClusterParameterGroup - Specifies the name of the cluster parameter group for the cluster.

$sel:masterUsername:DBCluster', dbCluster_masterUsername - Contains the master user name for the cluster.

$sel:dbClusterResourceId:DBCluster', dbCluster_dbClusterResourceId - The Region-unique, immutable identifier for the cluster. This identifier is found in CloudTrail log entries whenever the KMS key for the cluster is accessed.

$sel:earliestRestorableTime:DBCluster', dbCluster_earliestRestorableTime - The earliest time to which a database can be restored with point-in-time restore.

$sel:engine:DBCluster', dbCluster_engine - Provides the name of the database engine to be used for this cluster.

$sel:dbClusterArn:DBCluster', dbCluster_dbClusterArn - The Amazon Resource Name (ARN) for the cluster.

$sel:latestRestorableTime:DBCluster', dbCluster_latestRestorableTime - Specifies the latest time to which a database can be restored with point-in-time restore.

$sel:preferredMaintenanceWindow:DBCluster', dbCluster_preferredMaintenanceWindow - Specifies the weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

$sel:availabilityZones:DBCluster', dbCluster_availabilityZones - Provides the list of Amazon EC2 Availability Zones that instances in the cluster can be created in.

$sel:kmsKeyId:DBCluster', dbCluster_kmsKeyId - If StorageEncrypted is true, the KMS key identifier for the encrypted cluster.

$sel:preferredBackupWindow:DBCluster', dbCluster_preferredBackupWindow - Specifies the daily time range during which automated backups are created if automated backups are enabled, as determined by the BackupRetentionPeriod.

$sel:associatedRoles:DBCluster', dbCluster_associatedRoles - Provides a list of the Identity and Access Management (IAM) roles that are associated with the cluster. (IAM) roles that are associated with a cluster grant permission for the cluster to access other Amazon Web Services services on your behalf.

$sel:vpcSecurityGroups:DBCluster', dbCluster_vpcSecurityGroups - Provides a list of virtual private cloud (VPC) security groups that the cluster belongs to.

$sel:backupRetentionPeriod:DBCluster', dbCluster_backupRetentionPeriod - Specifies the number of days for which automatic snapshots are retained.

$sel:dbSubnetGroup:DBCluster', dbCluster_dbSubnetGroup - Specifies information on the subnet group that is associated with the cluster, including the name, description, and subnets in the subnet group.

$sel:multiAZ:DBCluster', dbCluster_multiAZ - Specifies whether the cluster has instances in multiple Availability Zones.

$sel:enabledCloudwatchLogsExports:DBCluster', dbCluster_enabledCloudwatchLogsExports - A list of log types that this cluster is configured to export to Amazon CloudWatch Logs.

$sel:clusterCreateTime:DBCluster', dbCluster_clusterCreateTime - Specifies the time when the cluster was created, in Universal Coordinated Time (UTC).

$sel:endpoint:DBCluster', dbCluster_endpoint - Specifies the connection endpoint for the primary instance of the cluster.

$sel:percentProgress:DBCluster', dbCluster_percentProgress - Specifies the progress of the operation as a percentage.

$sel:readerEndpoint:DBCluster', dbCluster_readerEndpoint - The reader endpoint for the cluster. The reader endpoint for a cluster load balances connections across the Amazon DocumentDB replicas that are available in a cluster. As clients request new connections to the reader endpoint, Amazon DocumentDB distributes the connection requests among the Amazon DocumentDB replicas in the cluster. This functionality can help balance your read workload across multiple Amazon DocumentDB replicas in your cluster.

If a failover occurs, and the Amazon DocumentDB replica that you are connected to is promoted to be the primary instance, your connection is dropped. To continue sending your read workload to other Amazon DocumentDB replicas in the cluster, you can then reconnect to the reader endpoint.

$sel:port:DBCluster', dbCluster_port - Specifies the port that the database engine is listening on.

dbCluster_engineVersion :: Lens' DBCluster (Maybe Text)

Indicates the database engine version.

dbCluster_status :: Lens' DBCluster (Maybe Text)

Specifies the current state of this cluster.

dbCluster_deletionProtection :: Lens' DBCluster (Maybe Bool)

Specifies whether this cluster can be deleted. If DeletionProtection is enabled, the cluster cannot be deleted unless it is modified and DeletionProtection is disabled. DeletionProtection protects clusters from being accidentally deleted.

dbCluster_storageEncrypted :: Lens' DBCluster (Maybe Bool)

Specifies whether the cluster is encrypted.

dbCluster_dbClusterIdentifier :: Lens' DBCluster (Maybe Text)

Contains a user-supplied cluster identifier. This identifier is the unique key that identifies a cluster.

dbCluster_dbClusterMembers :: Lens' DBCluster (Maybe [DBClusterMember])

Provides the list of instances that make up the cluster.

dbCluster_readReplicaIdentifiers :: Lens' DBCluster (Maybe [Text])

Contains one or more identifiers of the secondary clusters that are associated with this cluster.

dbCluster_replicationSourceIdentifier :: Lens' DBCluster (Maybe Text)

Contains the identifier of the source cluster if this cluster is a secondary cluster.

dbCluster_hostedZoneId :: Lens' DBCluster (Maybe Text)

Specifies the ID that Amazon Route 53 assigns when you create a hosted zone.

dbCluster_dbClusterParameterGroup :: Lens' DBCluster (Maybe Text)

Specifies the name of the cluster parameter group for the cluster.

dbCluster_masterUsername :: Lens' DBCluster (Maybe Text)

Contains the master user name for the cluster.

dbCluster_dbClusterResourceId :: Lens' DBCluster (Maybe Text)

The Region-unique, immutable identifier for the cluster. This identifier is found in CloudTrail log entries whenever the KMS key for the cluster is accessed.

dbCluster_earliestRestorableTime :: Lens' DBCluster (Maybe UTCTime)

The earliest time to which a database can be restored with point-in-time restore.

dbCluster_engine :: Lens' DBCluster (Maybe Text)

Provides the name of the database engine to be used for this cluster.

dbCluster_dbClusterArn :: Lens' DBCluster (Maybe Text)

The Amazon Resource Name (ARN) for the cluster.

dbCluster_latestRestorableTime :: Lens' DBCluster (Maybe UTCTime)

Specifies the latest time to which a database can be restored with point-in-time restore.

dbCluster_preferredMaintenanceWindow :: Lens' DBCluster (Maybe Text)

Specifies the weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

dbCluster_availabilityZones :: Lens' DBCluster (Maybe [Text])

Provides the list of Amazon EC2 Availability Zones that instances in the cluster can be created in.

dbCluster_kmsKeyId :: Lens' DBCluster (Maybe Text)

If StorageEncrypted is true, the KMS key identifier for the encrypted cluster.

dbCluster_preferredBackupWindow :: Lens' DBCluster (Maybe Text)

Specifies the daily time range during which automated backups are created if automated backups are enabled, as determined by the BackupRetentionPeriod.

dbCluster_associatedRoles :: Lens' DBCluster (Maybe [DBClusterRole])

Provides a list of the Identity and Access Management (IAM) roles that are associated with the cluster. (IAM) roles that are associated with a cluster grant permission for the cluster to access other Amazon Web Services services on your behalf.

dbCluster_vpcSecurityGroups :: Lens' DBCluster (Maybe [VpcSecurityGroupMembership])

Provides a list of virtual private cloud (VPC) security groups that the cluster belongs to.

dbCluster_backupRetentionPeriod :: Lens' DBCluster (Maybe Int)

Specifies the number of days for which automatic snapshots are retained.

dbCluster_dbSubnetGroup :: Lens' DBCluster (Maybe Text)

Specifies information on the subnet group that is associated with the cluster, including the name, description, and subnets in the subnet group.

dbCluster_multiAZ :: Lens' DBCluster (Maybe Bool)

Specifies whether the cluster has instances in multiple Availability Zones.

dbCluster_enabledCloudwatchLogsExports :: Lens' DBCluster (Maybe [Text])

A list of log types that this cluster is configured to export to Amazon CloudWatch Logs.

dbCluster_clusterCreateTime :: Lens' DBCluster (Maybe UTCTime)

Specifies the time when the cluster was created, in Universal Coordinated Time (UTC).

dbCluster_endpoint :: Lens' DBCluster (Maybe Text)

Specifies the connection endpoint for the primary instance of the cluster.

dbCluster_percentProgress :: Lens' DBCluster (Maybe Text)

Specifies the progress of the operation as a percentage.

dbCluster_readerEndpoint :: Lens' DBCluster (Maybe Text)

The reader endpoint for the cluster. The reader endpoint for a cluster load balances connections across the Amazon DocumentDB replicas that are available in a cluster. As clients request new connections to the reader endpoint, Amazon DocumentDB distributes the connection requests among the Amazon DocumentDB replicas in the cluster. This functionality can help balance your read workload across multiple Amazon DocumentDB replicas in your cluster.

If a failover occurs, and the Amazon DocumentDB replica that you are connected to is promoted to be the primary instance, your connection is dropped. To continue sending your read workload to other Amazon DocumentDB replicas in the cluster, you can then reconnect to the reader endpoint.

dbCluster_port :: Lens' DBCluster (Maybe Int)

Specifies the port that the database engine is listening on.