Network.AWS.DocumentDB.Types.DBClusterMember

Description

Documentation

data DBClusterMember

Contains information about an instance that is part of a cluster.

See: newDBClusterMember smart constructor.

Constructors

DBClusterMember'

Fields

promotionTier :: Maybe Int

A value that specifies the order in which an Amazon DocumentDB replica is promoted to the primary instance after a failure of the existing primary instance.

dbInstanceIdentifier :: Maybe Text

Specifies the instance identifier for this member of the cluster.

isClusterWriter :: Maybe Bool

A value that is true if the cluster member is the primary instance for the cluster and false otherwise.

dbClusterParameterGroupStatus :: Maybe Text

Specifies the status of the cluster parameter group for this member of the DB cluster.

newDBClusterMember :: DBClusterMember

Create a value of DBClusterMember with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:promotionTier:DBClusterMember', dbClusterMember_promotionTier - A value that specifies the order in which an Amazon DocumentDB replica is promoted to the primary instance after a failure of the existing primary instance.

$sel:dbInstanceIdentifier:DBClusterMember', dbClusterMember_dbInstanceIdentifier - Specifies the instance identifier for this member of the cluster.

$sel:isClusterWriter:DBClusterMember', dbClusterMember_isClusterWriter - A value that is true if the cluster member is the primary instance for the cluster and false otherwise.

$sel:dbClusterParameterGroupStatus:DBClusterMember', dbClusterMember_dbClusterParameterGroupStatus - Specifies the status of the cluster parameter group for this member of the DB cluster.

dbClusterMember_promotionTier :: Lens' DBClusterMember (Maybe Int)

A value that specifies the order in which an Amazon DocumentDB replica is promoted to the primary instance after a failure of the existing primary instance.

dbClusterMember_dbInstanceIdentifier :: Lens' DBClusterMember (Maybe Text)

Specifies the instance identifier for this member of the cluster.

dbClusterMember_isClusterWriter :: Lens' DBClusterMember (Maybe Bool)

A value that is true if the cluster member is the primary instance for the cluster and false otherwise.

dbClusterMember_dbClusterParameterGroupStatus :: Lens' DBClusterMember (Maybe Text)

Specifies the status of the cluster parameter group for this member of the DB cluster.