Network.AWS.DocumentDB.Types.DBInstanceStatusInfo

Description

Documentation

data DBInstanceStatusInfo

Provides a list of status information for an instance.

See: newDBInstanceStatusInfo smart constructor.

Constructors

DBInstanceStatusInfo'

Fields

status :: Maybe Text

Status of the instance. For a StatusType of read replica, the values can be replicating, error, stopped, or terminated.

normal :: Maybe Bool

A Boolean value that is true if the instance is operating normally, or false if the instance is in an error state.

statusType :: Maybe Text

This value is currently "read replication."

message :: Maybe Text

Details of the error if there is an error for the instance. If the instance is not in an error state, this value is blank.

newDBInstanceStatusInfo :: DBInstanceStatusInfo

Create a value of DBInstanceStatusInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DBInstanceStatusInfo', dbInstanceStatusInfo_status - Status of the instance. For a StatusType of read replica, the values can be replicating, error, stopped, or terminated.

$sel:normal:DBInstanceStatusInfo', dbInstanceStatusInfo_normal - A Boolean value that is true if the instance is operating normally, or false if the instance is in an error state.

$sel:statusType:DBInstanceStatusInfo', dbInstanceStatusInfo_statusType - This value is currently "read replication."

$sel:message:DBInstanceStatusInfo', dbInstanceStatusInfo_message - Details of the error if there is an error for the instance. If the instance is not in an error state, this value is blank.

dbInstanceStatusInfo_status :: Lens' DBInstanceStatusInfo (Maybe Text)

Status of the instance. For a StatusType of read replica, the values can be replicating, error, stopped, or terminated.

dbInstanceStatusInfo_normal :: Lens' DBInstanceStatusInfo (Maybe Bool)

A Boolean value that is true if the instance is operating normally, or false if the instance is in an error state.

dbInstanceStatusInfo_statusType :: Lens' DBInstanceStatusInfo (Maybe Text)

This value is currently "read replication."

dbInstanceStatusInfo_message :: Lens' DBInstanceStatusInfo (Maybe Text)

Details of the error if there is an error for the instance. If the instance is not in an error state, this value is blank.