Network.AWS.DocumentDB.Types.DBSubnetGroup

Description

Documentation

data DBSubnetGroup

Detailed information about a subnet group.

See: newDBSubnetGroup smart constructor.

Constructors

DBSubnetGroup'

Fields

dbSubnetGroupName :: Maybe Text

The name of the subnet group.

vpcId :: Maybe Text

Provides the virtual private cloud (VPC) ID of the subnet group.

subnets :: Maybe [Subnet]

Detailed information about one or more subnets within a subnet group.

dbSubnetGroupDescription :: Maybe Text

Provides the description of the subnet group.

dbSubnetGroupArn :: Maybe Text

The Amazon Resource Name (ARN) for the DB subnet group.

subnetGroupStatus :: Maybe Text

Provides the status of the subnet group.

newDBSubnetGroup :: DBSubnetGroup

Create a value of DBSubnetGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbSubnetGroupName:DBSubnetGroup', dbSubnetGroup_dbSubnetGroupName - The name of the subnet group.

$sel:vpcId:DBSubnetGroup', dbSubnetGroup_vpcId - Provides the virtual private cloud (VPC) ID of the subnet group.

$sel:subnets:DBSubnetGroup', dbSubnetGroup_subnets - Detailed information about one or more subnets within a subnet group.

$sel:dbSubnetGroupDescription:DBSubnetGroup', dbSubnetGroup_dbSubnetGroupDescription - Provides the description of the subnet group.

$sel:dbSubnetGroupArn:DBSubnetGroup', dbSubnetGroup_dbSubnetGroupArn - The Amazon Resource Name (ARN) for the DB subnet group.

$sel:subnetGroupStatus:DBSubnetGroup', dbSubnetGroup_subnetGroupStatus - Provides the status of the subnet group.

dbSubnetGroup_dbSubnetGroupName :: Lens' DBSubnetGroup (Maybe Text)

The name of the subnet group.

dbSubnetGroup_vpcId :: Lens' DBSubnetGroup (Maybe Text)

Provides the virtual private cloud (VPC) ID of the subnet group.

dbSubnetGroup_subnets :: Lens' DBSubnetGroup (Maybe [Subnet])

Detailed information about one or more subnets within a subnet group.

dbSubnetGroup_dbSubnetGroupDescription :: Lens' DBSubnetGroup (Maybe Text)

Provides the description of the subnet group.

dbSubnetGroup_dbSubnetGroupArn :: Lens' DBSubnetGroup (Maybe Text)

The Amazon Resource Name (ARN) for the DB subnet group.

dbSubnetGroup_subnetGroupStatus :: Lens' DBSubnetGroup (Maybe Text)

Provides the status of the subnet group.