Network.AWS.DynamoDB.DescribeContributorInsights

Description

Returns information about contributor insights, for a given table or global secondary index.

Creating a Request

data DescribeContributorInsights

See: newDescribeContributorInsights smart constructor.

Constructors

DescribeContributorInsights'

Fields

indexName :: Maybe Text

The name of the global secondary index to describe, if applicable.

tableName :: Text

The name of the table to describe.

newDescribeContributorInsights

Arguments

:: Text
$sel:tableName:DescribeContributorInsights'

-> DescribeContributorInsights

Create a value of DescribeContributorInsights with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:indexName:DescribeContributorInsights', describeContributorInsights_indexName - The name of the global secondary index to describe, if applicable.

$sel:tableName:DescribeContributorInsights', describeContributorInsights_tableName - The name of the table to describe.

Request Lenses

describeContributorInsights_indexName :: Lens' DescribeContributorInsights (Maybe Text)

The name of the global secondary index to describe, if applicable.

describeContributorInsights_tableName :: Lens' DescribeContributorInsights Text

The name of the table to describe.

Destructuring the Response

data DescribeContributorInsightsResponse

See: newDescribeContributorInsightsResponse smart constructor.

Constructors

DescribeContributorInsightsResponse'

Fields

tableName :: Maybe Text

The name of the table being described.

indexName :: Maybe Text

The name of the global secondary index being described.

contributorInsightsStatus :: Maybe ContributorInsightsStatus

Current Status contributor insights.

failureException :: Maybe FailureException

Returns information about the last failure that encountered.

The most common exceptions for a FAILED status are:

LimitExceededException - Per-account Amazon CloudWatch Contributor Insights rule limit reached. Please disable Contributor Insights for other tables/indexes OR disable Contributor Insights rules before retrying.

AccessDeniedException - Amazon CloudWatch Contributor Insights rules cannot be modified due to insufficient permissions.

AccessDeniedException - Failed to create service-linked role for Contributor Insights due to insufficient permissions.

InternalServerError - Failed to create Amazon CloudWatch Contributor Insights rules. Please retry request.

contributorInsightsRuleList :: Maybe [Text]

List of names of the associated Alpine rules.

lastUpdateDateTime :: Maybe POSIX

Timestamp of the last time the status was changed.

httpStatus :: Int

The response's http status code.

newDescribeContributorInsightsResponse

Arguments

:: Int
$sel:httpStatus:DescribeContributorInsightsResponse'

-> DescribeContributorInsightsResponse

Create a value of DescribeContributorInsightsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tableName:DescribeContributorInsights', describeContributorInsightsResponse_tableName - The name of the table being described.

$sel:indexName:DescribeContributorInsights', describeContributorInsightsResponse_indexName - The name of the global secondary index being described.

$sel:contributorInsightsStatus:DescribeContributorInsightsResponse', describeContributorInsightsResponse_contributorInsightsStatus - Current Status contributor insights.

$sel:failureException:DescribeContributorInsightsResponse', describeContributorInsightsResponse_failureException - Returns information about the last failure that encountered.

The most common exceptions for a FAILED status are:

LimitExceededException - Per-account Amazon CloudWatch Contributor Insights rule limit reached. Please disable Contributor Insights for other tables/indexes OR disable Contributor Insights rules before retrying.

AccessDeniedException - Amazon CloudWatch Contributor Insights rules cannot be modified due to insufficient permissions.

AccessDeniedException - Failed to create service-linked role for Contributor Insights due to insufficient permissions.

InternalServerError - Failed to create Amazon CloudWatch Contributor Insights rules. Please retry request.

$sel:contributorInsightsRuleList:DescribeContributorInsightsResponse', describeContributorInsightsResponse_contributorInsightsRuleList - List of names of the associated Alpine rules.

$sel:lastUpdateDateTime:DescribeContributorInsightsResponse', describeContributorInsightsResponse_lastUpdateDateTime - Timestamp of the last time the status was changed.

$sel:httpStatus:DescribeContributorInsightsResponse', describeContributorInsightsResponse_httpStatus - The response's http status code.

Response Lenses

describeContributorInsightsResponse_tableName :: Lens' DescribeContributorInsightsResponse (Maybe Text)

The name of the table being described.

describeContributorInsightsResponse_indexName :: Lens' DescribeContributorInsightsResponse (Maybe Text)

The name of the global secondary index being described.

describeContributorInsightsResponse_contributorInsightsStatus :: Lens' DescribeContributorInsightsResponse (Maybe ContributorInsightsStatus)

Current Status contributor insights.

describeContributorInsightsResponse_failureException :: Lens' DescribeContributorInsightsResponse (Maybe FailureException)

Returns information about the last failure that encountered.

The most common exceptions for a FAILED status are:

LimitExceededException - Per-account Amazon CloudWatch Contributor Insights rule limit reached. Please disable Contributor Insights for other tables/indexes OR disable Contributor Insights rules before retrying.

AccessDeniedException - Amazon CloudWatch Contributor Insights rules cannot be modified due to insufficient permissions.

AccessDeniedException - Failed to create service-linked role for Contributor Insights due to insufficient permissions.

InternalServerError - Failed to create Amazon CloudWatch Contributor Insights rules. Please retry request.

describeContributorInsightsResponse_contributorInsightsRuleList :: Lens' DescribeContributorInsightsResponse (Maybe [Text])

List of names of the associated Alpine rules.

describeContributorInsightsResponse_lastUpdateDateTime :: Lens' DescribeContributorInsightsResponse (Maybe UTCTime)

Timestamp of the last time the status was changed.

describeContributorInsightsResponse_httpStatus :: Lens' DescribeContributorInsightsResponse Int

The response's http status code.