Network.AWS.DynamoDB.DescribeExport

Description

Describes an existing table export.

Creating a Request

data DescribeExport

See: newDescribeExport smart constructor.

Constructors

DescribeExport'

Fields

exportArn :: Text

The Amazon Resource Name (ARN) associated with the export.

newDescribeExport

Arguments

:: Text
$sel:exportArn:DescribeExport'

-> DescribeExport

Create a value of DescribeExport with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:exportArn:DescribeExport', describeExport_exportArn - The Amazon Resource Name (ARN) associated with the export.

Request Lenses

describeExport_exportArn :: Lens' DescribeExport Text

The Amazon Resource Name (ARN) associated with the export.

Destructuring the Response

data DescribeExportResponse

See: newDescribeExportResponse smart constructor.

Constructors

DescribeExportResponse'

Fields

exportDescription :: Maybe ExportDescription

Represents the properties of the export.

httpStatus :: Int

The response's http status code.

newDescribeExportResponse

Arguments

:: Int
$sel:httpStatus:DescribeExportResponse'

-> DescribeExportResponse

Create a value of DescribeExportResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:exportDescription:DescribeExportResponse', describeExportResponse_exportDescription - Represents the properties of the export.

$sel:httpStatus:DescribeExportResponse', describeExportResponse_httpStatus - The response's http status code.

Response Lenses

describeExportResponse_exportDescription :: Lens' DescribeExportResponse (Maybe ExportDescription)

Represents the properties of the export.

describeExportResponse_httpStatus :: Lens' DescribeExportResponse Int

The response's http status code.