Network.AWS.DynamoDB.DescribeTimeToLive

Description

Gives a description of the Time to Live (TTL) status on the specified table.

Creating a Request

data DescribeTimeToLive

See: newDescribeTimeToLive smart constructor.

Constructors

DescribeTimeToLive'

Fields

tableName :: Text

The name of the table to be described.

newDescribeTimeToLive

Arguments

:: Text
$sel:tableName:DescribeTimeToLive'

-> DescribeTimeToLive

Create a value of DescribeTimeToLive with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tableName:DescribeTimeToLive', describeTimeToLive_tableName - The name of the table to be described.

Request Lenses

describeTimeToLive_tableName :: Lens' DescribeTimeToLive Text

The name of the table to be described.

Destructuring the Response

data DescribeTimeToLiveResponse

See: newDescribeTimeToLiveResponse smart constructor.

Constructors

DescribeTimeToLiveResponse'

Fields

timeToLiveDescription :: Maybe TimeToLiveDescription

httpStatus :: Int

The response's http status code.

newDescribeTimeToLiveResponse

Arguments

:: Int
$sel:httpStatus:DescribeTimeToLiveResponse'

-> DescribeTimeToLiveResponse

Create a value of DescribeTimeToLiveResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timeToLiveDescription:DescribeTimeToLiveResponse', describeTimeToLiveResponse_timeToLiveDescription -

$sel:httpStatus:DescribeTimeToLiveResponse', describeTimeToLiveResponse_httpStatus - The response's http status code.

Response Lenses

describeTimeToLiveResponse_timeToLiveDescription :: Lens' DescribeTimeToLiveResponse (Maybe TimeToLiveDescription)

describeTimeToLiveResponse_httpStatus :: Lens' DescribeTimeToLiveResponse Int

The response's http status code.