Network.AWS.DynamoDB.EnableKinesisStreamingDestination

Description

Starts table data replication to the specified Kinesis data stream at a timestamp chosen during the enable workflow. If this operation doesn't return results immediately, use DescribeKinesisStreamingDestination to check if streaming to the Kinesis data stream is ACTIVE.

Creating a Request

data EnableKinesisStreamingDestination

See: newEnableKinesisStreamingDestination smart constructor.

Constructors

EnableKinesisStreamingDestination'

Fields

tableName :: Text

The name of the DynamoDB table.

streamArn :: Text

The ARN for a Kinesis data stream.

newEnableKinesisStreamingDestination

Arguments

:: Text
$sel:tableName:EnableKinesisStreamingDestination'

-> Text
$sel:streamArn:EnableKinesisStreamingDestination'

-> EnableKinesisStreamingDestination

Create a value of EnableKinesisStreamingDestination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tableName:EnableKinesisStreamingDestination', enableKinesisStreamingDestination_tableName - The name of the DynamoDB table.

$sel:streamArn:EnableKinesisStreamingDestination', enableKinesisStreamingDestination_streamArn - The ARN for a Kinesis data stream.

Request Lenses

enableKinesisStreamingDestination_tableName :: Lens' EnableKinesisStreamingDestination Text

The name of the DynamoDB table.

enableKinesisStreamingDestination_streamArn :: Lens' EnableKinesisStreamingDestination Text

The ARN for a Kinesis data stream.

Destructuring the Response

data KinesisStreamingDestinationOutput

See: newKinesisStreamingDestinationOutput smart constructor.

Constructors

KinesisStreamingDestinationOutput'

Fields

tableName :: Maybe Text

The name of the table being modified.

streamArn :: Maybe Text

The ARN for the specific Kinesis data stream.

destinationStatus :: Maybe DestinationStatus

The current status of the replication.

newKinesisStreamingDestinationOutput :: KinesisStreamingDestinationOutput

Create a value of KinesisStreamingDestinationOutput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tableName:KinesisStreamingDestinationOutput', kinesisStreamingDestinationOutput_tableName - The name of the table being modified.

$sel:streamArn:KinesisStreamingDestinationOutput', kinesisStreamingDestinationOutput_streamArn - The ARN for the specific Kinesis data stream.

$sel:destinationStatus:KinesisStreamingDestinationOutput', kinesisStreamingDestinationOutput_destinationStatus - The current status of the replication.

Response Lenses

kinesisStreamingDestinationOutput_tableName :: Lens' KinesisStreamingDestinationOutput (Maybe Text)

The name of the table being modified.

kinesisStreamingDestinationOutput_streamArn :: Lens' KinesisStreamingDestinationOutput (Maybe Text)

The ARN for the specific Kinesis data stream.

kinesisStreamingDestinationOutput_destinationStatus :: Lens' KinesisStreamingDestinationOutput (Maybe DestinationStatus)

The current status of the replication.