Network.AWS.DynamoDB.ExecuteStatement

Description

This operation allows you to perform reads and singleton writes on data stored in DynamoDB, using PartiQL.

Creating a Request

data ExecuteStatement

See: newExecuteStatement smart constructor.

Constructors

ExecuteStatement'

Fields

nextToken :: Maybe Text

Set this value to get remaining results, if NextToken was returned in the statement response.

consistentRead :: Maybe Bool

The consistency of a read operation. If set to true, then a strongly consistent read is used; otherwise, an eventually consistent read is used.

parameters :: Maybe (NonEmpty AttributeValue)

The parameters for the PartiQL statement, if any.

statement :: Text

The PartiQL statement representing the operation to run.

newExecuteStatement

Arguments

:: Text
$sel:statement:ExecuteStatement'

-> ExecuteStatement

Create a value of ExecuteStatement with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ExecuteStatement', executeStatement_nextToken - Set this value to get remaining results, if NextToken was returned in the statement response.

$sel:consistentRead:ExecuteStatement', executeStatement_consistentRead - The consistency of a read operation. If set to true, then a strongly consistent read is used; otherwise, an eventually consistent read is used.

$sel:parameters:ExecuteStatement', executeStatement_parameters - The parameters for the PartiQL statement, if any.

$sel:statement:ExecuteStatement', executeStatement_statement - The PartiQL statement representing the operation to run.

Request Lenses

executeStatement_nextToken :: Lens' ExecuteStatement (Maybe Text)

Set this value to get remaining results, if NextToken was returned in the statement response.

executeStatement_consistentRead :: Lens' ExecuteStatement (Maybe Bool)

The consistency of a read operation. If set to true, then a strongly consistent read is used; otherwise, an eventually consistent read is used.

executeStatement_parameters :: Lens' ExecuteStatement (Maybe (NonEmpty AttributeValue))

The parameters for the PartiQL statement, if any.

executeStatement_statement :: Lens' ExecuteStatement Text

The PartiQL statement representing the operation to run.

Destructuring the Response

data ExecuteStatementResponse

See: newExecuteStatementResponse smart constructor.

Constructors

ExecuteStatementResponse'

Fields

nextToken :: Maybe Text

If the response of a read request exceeds the response payload limit DynamoDB will set this value in the response. If set, you can use that this value in the subsequent request to get the remaining results.

items :: Maybe [HashMap Text AttributeValue]

If a read operation was used, this property will contain the result of the reade operation; a map of attribute names and their values. For the write operations this value will be empty.

httpStatus :: Int

The response's http status code.

newExecuteStatementResponse

Arguments

:: Int
$sel:httpStatus:ExecuteStatementResponse'

-> ExecuteStatementResponse

Create a value of ExecuteStatementResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ExecuteStatement', executeStatementResponse_nextToken - If the response of a read request exceeds the response payload limit DynamoDB will set this value in the response. If set, you can use that this value in the subsequent request to get the remaining results.

$sel:items:ExecuteStatementResponse', executeStatementResponse_items - If a read operation was used, this property will contain the result of the reade operation; a map of attribute names and their values. For the write operations this value will be empty.

$sel:httpStatus:ExecuteStatementResponse', executeStatementResponse_httpStatus - The response's http status code.

Response Lenses

executeStatementResponse_nextToken :: Lens' ExecuteStatementResponse (Maybe Text)

If the response of a read request exceeds the response payload limit DynamoDB will set this value in the response. If set, you can use that this value in the subsequent request to get the remaining results.

executeStatementResponse_items :: Lens' ExecuteStatementResponse (Maybe [HashMap Text AttributeValue])

If a read operation was used, this property will contain the result of the reade operation; a map of attribute names and their values. For the write operations this value will be empty.

executeStatementResponse_httpStatus :: Lens' ExecuteStatementResponse Int

The response's http status code.