Network.AWS.DynamoDB.ListBackups

Description

List backups associated with an AWS account. To list backups for a given table, specify TableName. ListBackups returns a paginated list of results with at most 1 MB worth of items in a page. You can also specify a maximum number of entries to be returned in a page.

In the request, start time is inclusive, but end time is exclusive. Note that these boundaries are for the time at which the original backup was requested.

You can call ListBackups a maximum of five times per second.

This operation returns paginated results.

Creating a Request

data ListBackups

See: newListBackups smart constructor.

Constructors

ListBackups'

Fields

tableName :: Maybe Text

The backups from the table specified by TableName are listed.

backupType :: Maybe BackupTypeFilter

The backups from the table specified by BackupType are listed.

Where BackupType can be:

USER - On-demand backup created by you.

SYSTEM - On-demand backup automatically created by DynamoDB.

ALL - All types of on-demand backups (USER and SYSTEM).

timeRangeLowerBound :: Maybe POSIX

Only backups created after this time are listed. TimeRangeLowerBound is inclusive.

exclusiveStartBackupArn :: Maybe Text

LastEvaluatedBackupArn is the Amazon Resource Name (ARN) of the backup last evaluated when the current page of results was returned, inclusive of the current page of results. This value may be specified as the ExclusiveStartBackupArn of a new ListBackups operation in order to fetch the next page of results.

limit :: Maybe Natural

Maximum number of backups to return at once.

timeRangeUpperBound :: Maybe POSIX

Only backups created before this time are listed. TimeRangeUpperBound is exclusive.

newListBackups :: ListBackups

Create a value of ListBackups with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tableName:ListBackups', listBackups_tableName - The backups from the table specified by TableName are listed.

$sel:backupType:ListBackups', listBackups_backupType - The backups from the table specified by BackupType are listed.

Where BackupType can be:

USER - On-demand backup created by you.

SYSTEM - On-demand backup automatically created by DynamoDB.

ALL - All types of on-demand backups (USER and SYSTEM).

$sel:timeRangeLowerBound:ListBackups', listBackups_timeRangeLowerBound - Only backups created after this time are listed. TimeRangeLowerBound is inclusive.

$sel:exclusiveStartBackupArn:ListBackups', listBackups_exclusiveStartBackupArn - LastEvaluatedBackupArn is the Amazon Resource Name (ARN) of the backup last evaluated when the current page of results was returned, inclusive of the current page of results. This value may be specified as the ExclusiveStartBackupArn of a new ListBackups operation in order to fetch the next page of results.

$sel:limit:ListBackups', listBackups_limit - Maximum number of backups to return at once.

$sel:timeRangeUpperBound:ListBackups', listBackups_timeRangeUpperBound - Only backups created before this time are listed. TimeRangeUpperBound is exclusive.

Request Lenses

listBackups_tableName :: Lens' ListBackups (Maybe Text)

The backups from the table specified by TableName are listed.

listBackups_backupType :: Lens' ListBackups (Maybe BackupTypeFilter)

The backups from the table specified by BackupType are listed.

Where BackupType can be:

USER - On-demand backup created by you.

SYSTEM - On-demand backup automatically created by DynamoDB.

ALL - All types of on-demand backups (USER and SYSTEM).

listBackups_timeRangeLowerBound :: Lens' ListBackups (Maybe UTCTime)

Only backups created after this time are listed. TimeRangeLowerBound is inclusive.

listBackups_exclusiveStartBackupArn :: Lens' ListBackups (Maybe Text)

LastEvaluatedBackupArn is the Amazon Resource Name (ARN) of the backup last evaluated when the current page of results was returned, inclusive of the current page of results. This value may be specified as the ExclusiveStartBackupArn of a new ListBackups operation in order to fetch the next page of results.

listBackups_limit :: Lens' ListBackups (Maybe Natural)

Maximum number of backups to return at once.

listBackups_timeRangeUpperBound :: Lens' ListBackups (Maybe UTCTime)

Only backups created before this time are listed. TimeRangeUpperBound is exclusive.

Destructuring the Response

data ListBackupsResponse

See: newListBackupsResponse smart constructor.

Constructors

ListBackupsResponse'

Fields

lastEvaluatedBackupArn :: Maybe Text

The ARN of the backup last evaluated when the current page of results was returned, inclusive of the current page of results. This value may be specified as the ExclusiveStartBackupArn of a new ListBackups operation in order to fetch the next page of results.

If LastEvaluatedBackupArn is empty, then the last page of results has been processed and there are no more results to be retrieved.

If LastEvaluatedBackupArn is not empty, this may or may not indicate that there is more data to be returned. All results are guaranteed to have been returned if and only if no value for LastEvaluatedBackupArn is returned.

backupSummaries :: Maybe [BackupSummary]

List of BackupSummary objects.

httpStatus :: Int

The response's http status code.

newListBackupsResponse

Arguments

:: Int
$sel:httpStatus:ListBackupsResponse'

-> ListBackupsResponse

Create a value of ListBackupsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastEvaluatedBackupArn:ListBackupsResponse', listBackupsResponse_lastEvaluatedBackupArn - The ARN of the backup last evaluated when the current page of results was returned, inclusive of the current page of results. This value may be specified as the ExclusiveStartBackupArn of a new ListBackups operation in order to fetch the next page of results.

If LastEvaluatedBackupArn is empty, then the last page of results has been processed and there are no more results to be retrieved.

If LastEvaluatedBackupArn is not empty, this may or may not indicate that there is more data to be returned. All results are guaranteed to have been returned if and only if no value for LastEvaluatedBackupArn is returned.

$sel:backupSummaries:ListBackupsResponse', listBackupsResponse_backupSummaries - List of BackupSummary objects.

$sel:httpStatus:ListBackupsResponse', listBackupsResponse_httpStatus - The response's http status code.

Response Lenses

listBackupsResponse_lastEvaluatedBackupArn :: Lens' ListBackupsResponse (Maybe Text)

The ARN of the backup last evaluated when the current page of results was returned, inclusive of the current page of results. This value may be specified as the ExclusiveStartBackupArn of a new ListBackups operation in order to fetch the next page of results.

If LastEvaluatedBackupArn is empty, then the last page of results has been processed and there are no more results to be retrieved.

If LastEvaluatedBackupArn is not empty, this may or may not indicate that there is more data to be returned. All results are guaranteed to have been returned if and only if no value for LastEvaluatedBackupArn is returned.

listBackupsResponse_backupSummaries :: Lens' ListBackupsResponse (Maybe [BackupSummary])

List of BackupSummary objects.

listBackupsResponse_httpStatus :: Lens' ListBackupsResponse Int

The response's http status code.