Network.AWS.DynamoDB.ListGlobalTables

Description

Lists all global tables that have a replica in the specified Region.

This operation only applies to Version 2017.11.29 of global tables.

Creating a Request

data ListGlobalTables

See: newListGlobalTables smart constructor.

Constructors

ListGlobalTables'

Fields

regionName :: Maybe Text

Lists the global tables in a specific Region.

exclusiveStartGlobalTableName :: Maybe Text

The first global table name that this operation will evaluate.

limit :: Maybe Natural

The maximum number of table names to return, if the parameter is not specified DynamoDB defaults to 100.

If the number of global tables DynamoDB finds reaches this limit, it stops the operation and returns the table names collected up to that point, with a table name in the LastEvaluatedGlobalTableName to apply in a subsequent operation to the ExclusiveStartGlobalTableName parameter.

newListGlobalTables :: ListGlobalTables

Create a value of ListGlobalTables with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:regionName:ListGlobalTables', listGlobalTables_regionName - Lists the global tables in a specific Region.

$sel:exclusiveStartGlobalTableName:ListGlobalTables', listGlobalTables_exclusiveStartGlobalTableName - The first global table name that this operation will evaluate.

$sel:limit:ListGlobalTables', listGlobalTables_limit - The maximum number of table names to return, if the parameter is not specified DynamoDB defaults to 100.

If the number of global tables DynamoDB finds reaches this limit, it stops the operation and returns the table names collected up to that point, with a table name in the LastEvaluatedGlobalTableName to apply in a subsequent operation to the ExclusiveStartGlobalTableName parameter.

Request Lenses

listGlobalTables_regionName :: Lens' ListGlobalTables (Maybe Text)

Lists the global tables in a specific Region.

listGlobalTables_exclusiveStartGlobalTableName :: Lens' ListGlobalTables (Maybe Text)

The first global table name that this operation will evaluate.

listGlobalTables_limit :: Lens' ListGlobalTables (Maybe Natural)

The maximum number of table names to return, if the parameter is not specified DynamoDB defaults to 100.

If the number of global tables DynamoDB finds reaches this limit, it stops the operation and returns the table names collected up to that point, with a table name in the LastEvaluatedGlobalTableName to apply in a subsequent operation to the ExclusiveStartGlobalTableName parameter.

Destructuring the Response

data ListGlobalTablesResponse

See: newListGlobalTablesResponse smart constructor.

Constructors

ListGlobalTablesResponse'

Fields

lastEvaluatedGlobalTableName :: Maybe Text

Last evaluated global table name.

globalTables :: Maybe [GlobalTable]

List of global table names.

httpStatus :: Int

The response's http status code.

newListGlobalTablesResponse

Arguments

:: Int
$sel:httpStatus:ListGlobalTablesResponse'

-> ListGlobalTablesResponse

Create a value of ListGlobalTablesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastEvaluatedGlobalTableName:ListGlobalTablesResponse', listGlobalTablesResponse_lastEvaluatedGlobalTableName - Last evaluated global table name.

$sel:globalTables:ListGlobalTablesResponse', listGlobalTablesResponse_globalTables - List of global table names.

$sel:httpStatus:ListGlobalTablesResponse', listGlobalTablesResponse_httpStatus - The response's http status code.

Response Lenses

listGlobalTablesResponse_lastEvaluatedGlobalTableName :: Lens' ListGlobalTablesResponse (Maybe Text)

Last evaluated global table name.

listGlobalTablesResponse_globalTables :: Lens' ListGlobalTablesResponse (Maybe [GlobalTable])

List of global table names.

listGlobalTablesResponse_httpStatus :: Lens' ListGlobalTablesResponse Int

The response's http status code.