Network.AWS.DynamoDB.ListTables

Description

Returns an array of table names associated with the current account and endpoint. The output from ListTables is paginated, with each page returning a maximum of 100 table names.

This operation returns paginated results.

Creating a Request

data ListTables

Represents the input of a ListTables operation.

See: newListTables smart constructor.

Constructors

ListTables'

Fields

exclusiveStartTableName :: Maybe Text

The first table name that this operation will evaluate. Use the value that was returned for LastEvaluatedTableName in a previous operation, so that you can obtain the next page of results.

limit :: Maybe Natural

A maximum number of table names to return. If this parameter is not specified, the limit is 100.

newListTables :: ListTables

Create a value of ListTables with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:exclusiveStartTableName:ListTables', listTables_exclusiveStartTableName - The first table name that this operation will evaluate. Use the value that was returned for LastEvaluatedTableName in a previous operation, so that you can obtain the next page of results.

$sel:limit:ListTables', listTables_limit - A maximum number of table names to return. If this parameter is not specified, the limit is 100.

Request Lenses

listTables_exclusiveStartTableName :: Lens' ListTables (Maybe Text)

The first table name that this operation will evaluate. Use the value that was returned for LastEvaluatedTableName in a previous operation, so that you can obtain the next page of results.

listTables_limit :: Lens' ListTables (Maybe Natural)

A maximum number of table names to return. If this parameter is not specified, the limit is 100.

Destructuring the Response

data ListTablesResponse

Represents the output of a ListTables operation.

See: newListTablesResponse smart constructor.

Constructors

ListTablesResponse'

Fields

lastEvaluatedTableName :: Maybe Text

The name of the last table in the current page of results. Use this value as the ExclusiveStartTableName in a new request to obtain the next page of results, until all the table names are returned.

If you do not receive a LastEvaluatedTableName value in the response, this means that there are no more table names to be retrieved.

tableNames :: Maybe [Text]

The names of the tables associated with the current account at the current endpoint. The maximum size of this array is 100.

If LastEvaluatedTableName also appears in the output, you can use this value as the ExclusiveStartTableName parameter in a subsequent ListTables request and obtain the next page of results.

httpStatus :: Int

The response's http status code.

newListTablesResponse

Arguments

:: Int
$sel:httpStatus:ListTablesResponse'

-> ListTablesResponse

Create a value of ListTablesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastEvaluatedTableName:ListTablesResponse', listTablesResponse_lastEvaluatedTableName - The name of the last table in the current page of results. Use this value as the ExclusiveStartTableName in a new request to obtain the next page of results, until all the table names are returned.

If you do not receive a LastEvaluatedTableName value in the response, this means that there are no more table names to be retrieved.

$sel:tableNames:ListTablesResponse', listTablesResponse_tableNames - The names of the tables associated with the current account at the current endpoint. The maximum size of this array is 100.

If LastEvaluatedTableName also appears in the output, you can use this value as the ExclusiveStartTableName parameter in a subsequent ListTables request and obtain the next page of results.

$sel:httpStatus:ListTablesResponse', listTablesResponse_httpStatus - The response's http status code.

Response Lenses

listTablesResponse_lastEvaluatedTableName :: Lens' ListTablesResponse (Maybe Text)

The name of the last table in the current page of results. Use this value as the ExclusiveStartTableName in a new request to obtain the next page of results, until all the table names are returned.

If you do not receive a LastEvaluatedTableName value in the response, this means that there are no more table names to be retrieved.

listTablesResponse_tableNames :: Lens' ListTablesResponse (Maybe [Text])

The names of the tables associated with the current account at the current endpoint. The maximum size of this array is 100.

If LastEvaluatedTableName also appears in the output, you can use this value as the ExclusiveStartTableName parameter in a subsequent ListTables request and obtain the next page of results.

listTablesResponse_httpStatus :: Lens' ListTablesResponse Int

The response's http status code.