Network.AWS.DynamoDB.TransactGetItems

Description

TransactGetItems is a synchronous operation that atomically retrieves multiple items from one or more tables (but not from indexes) in a single account and Region. A TransactGetItems call can contain up to 25 TransactGetItem objects, each of which contains a Get structure that specifies an item to retrieve from a table in the account and Region. A call to TransactGetItems cannot retrieve items from tables in more than one AWS account or Region. The aggregate size of the items in the transaction cannot exceed 4 MB.

DynamoDB rejects the entire TransactGetItems request if any of the following is true:

A conflicting operation is in the process of updating an item to be read.

There is insufficient provisioned capacity for the transaction to be completed.

There is a user error, such as an invalid data format.

The aggregate size of the items in the transaction cannot exceed 4 MB.

Creating a Request

data TransactGetItems

See: newTransactGetItems smart constructor.

Constructors

TransactGetItems'

Fields

returnConsumedCapacity :: Maybe ReturnConsumedCapacity

A value of TOTAL causes consumed capacity information to be returned, and a value of NONE prevents that information from being returned. No other value is valid.

transactItems :: NonEmpty TransactGetItem

An ordered array of up to 25 TransactGetItem objects, each of which contains a Get structure.

newTransactGetItems

Arguments

:: NonEmpty TransactGetItem
$sel:transactItems:TransactGetItems'

-> TransactGetItems

Create a value of TransactGetItems with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:returnConsumedCapacity:TransactGetItems', transactGetItems_returnConsumedCapacity - A value of TOTAL causes consumed capacity information to be returned, and a value of NONE prevents that information from being returned. No other value is valid.

$sel:transactItems:TransactGetItems', transactGetItems_transactItems - An ordered array of up to 25 TransactGetItem objects, each of which contains a Get structure.

Request Lenses

transactGetItems_returnConsumedCapacity :: Lens' TransactGetItems (Maybe ReturnConsumedCapacity)

A value of TOTAL causes consumed capacity information to be returned, and a value of NONE prevents that information from being returned. No other value is valid.

transactGetItems_transactItems :: Lens' TransactGetItems (NonEmpty TransactGetItem)

An ordered array of up to 25 TransactGetItem objects, each of which contains a Get structure.

Destructuring the Response

data TransactGetItemsResponse

See: newTransactGetItemsResponse smart constructor.

Constructors

TransactGetItemsResponse'

Fields

consumedCapacity :: Maybe [ConsumedCapacity]

If the ReturnConsumedCapacity value was TOTAL, this is an array of ConsumedCapacity objects, one for each table addressed by TransactGetItem objects in the TransactItems parameter. These ConsumedCapacity objects report the read-capacity units consumed by the TransactGetItems call in that table.

responses :: Maybe (NonEmpty ItemResponse)

An ordered array of up to 25 ItemResponse objects, each of which corresponds to the TransactGetItem object in the same position in the TransactItems array. Each ItemResponse object contains a Map of the name-value pairs that are the projected attributes of the requested item.

If a requested item could not be retrieved, the corresponding ItemResponse object is Null, or if the requested item has no projected attributes, the corresponding ItemResponse object is an empty Map.

httpStatus :: Int

The response's http status code.

newTransactGetItemsResponse

Arguments

:: Int
$sel:httpStatus:TransactGetItemsResponse'

-> TransactGetItemsResponse

Create a value of TransactGetItemsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:consumedCapacity:TransactGetItemsResponse', transactGetItemsResponse_consumedCapacity - If the ReturnConsumedCapacity value was TOTAL, this is an array of ConsumedCapacity objects, one for each table addressed by TransactGetItem objects in the TransactItems parameter. These ConsumedCapacity objects report the read-capacity units consumed by the TransactGetItems call in that table.

$sel:responses:TransactGetItemsResponse', transactGetItemsResponse_responses - An ordered array of up to 25 ItemResponse objects, each of which corresponds to the TransactGetItem object in the same position in the TransactItems array. Each ItemResponse object contains a Map of the name-value pairs that are the projected attributes of the requested item.

If a requested item could not be retrieved, the corresponding ItemResponse object is Null, or if the requested item has no projected attributes, the corresponding ItemResponse object is an empty Map.

$sel:httpStatus:TransactGetItemsResponse', transactGetItemsResponse_httpStatus - The response's http status code.

Response Lenses

transactGetItemsResponse_consumedCapacity :: Lens' TransactGetItemsResponse (Maybe [ConsumedCapacity])

If the ReturnConsumedCapacity value was TOTAL, this is an array of ConsumedCapacity objects, one for each table addressed by TransactGetItem objects in the TransactItems parameter. These ConsumedCapacity objects report the read-capacity units consumed by the TransactGetItems call in that table.

transactGetItemsResponse_responses :: Lens' TransactGetItemsResponse (Maybe (NonEmpty ItemResponse))

An ordered array of up to 25 ItemResponse objects, each of which corresponds to the TransactGetItem object in the same position in the TransactItems array. Each ItemResponse object contains a Map of the name-value pairs that are the projected attributes of the requested item.

If a requested item could not be retrieved, the corresponding ItemResponse object is Null, or if the requested item has no projected attributes, the corresponding ItemResponse object is an empty Map.

transactGetItemsResponse_httpStatus :: Lens' TransactGetItemsResponse Int

The response's http status code.