Network.AWS.DynamoDB.Types.ArchivalSummary

Description

Documentation

data ArchivalSummary

Contains details of a table archival operation.

See: newArchivalSummary smart constructor.

Constructors

ArchivalSummary'

Fields

archivalReason :: Maybe Text

The reason DynamoDB archived the table. Currently, the only possible value is:

INACCESSIBLE_ENCRYPTION_CREDENTIALS - The table was archived due to the table's AWS KMS key being inaccessible for more than seven days. An On-Demand backup was created at the archival time.

archivalDateTime :: Maybe POSIX

The date and time when table archival was initiated by DynamoDB, in UNIX epoch time format.

archivalBackupArn :: Maybe Text

The Amazon Resource Name (ARN) of the backup the table was archived to, when applicable in the archival reason. If you wish to restore this backup to the same table name, you will need to delete the original table.

newArchivalSummary :: ArchivalSummary

Create a value of ArchivalSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:archivalReason:ArchivalSummary', archivalSummary_archivalReason - The reason DynamoDB archived the table. Currently, the only possible value is:

INACCESSIBLE_ENCRYPTION_CREDENTIALS - The table was archived due to the table's AWS KMS key being inaccessible for more than seven days. An On-Demand backup was created at the archival time.

$sel:archivalDateTime:ArchivalSummary', archivalSummary_archivalDateTime - The date and time when table archival was initiated by DynamoDB, in UNIX epoch time format.

$sel:archivalBackupArn:ArchivalSummary', archivalSummary_archivalBackupArn - The Amazon Resource Name (ARN) of the backup the table was archived to, when applicable in the archival reason. If you wish to restore this backup to the same table name, you will need to delete the original table.

archivalSummary_archivalReason :: Lens' ArchivalSummary (Maybe Text)

The reason DynamoDB archived the table. Currently, the only possible value is:

INACCESSIBLE_ENCRYPTION_CREDENTIALS - The table was archived due to the table's AWS KMS key being inaccessible for more than seven days. An On-Demand backup was created at the archival time.

archivalSummary_archivalDateTime :: Lens' ArchivalSummary (Maybe UTCTime)

The date and time when table archival was initiated by DynamoDB, in UNIX epoch time format.

archivalSummary_archivalBackupArn :: Lens' ArchivalSummary (Maybe Text)

The Amazon Resource Name (ARN) of the backup the table was archived to, when applicable in the archival reason. If you wish to restore this backup to the same table name, you will need to delete the original table.