Network.AWS.DynamoDB.Types.AutoScalingTargetTrackingScalingPolicyConfigurationUpdate

Description

Documentation

data AutoScalingTargetTrackingScalingPolicyConfigurationUpdate

Represents the settings of a target tracking scaling policy that will be modified.

See: newAutoScalingTargetTrackingScalingPolicyConfigurationUpdate smart constructor.

Constructors

AutoScalingTargetTrackingScalingPolicyConfigurationUpdate'

Fields

scaleInCooldown :: Maybe Int

The amount of time, in seconds, after a scale in activity completes before another scale in activity can start. The cooldown period is used to block subsequent scale in requests until it has expired. You should scale in conservatively to protect your application's availability. However, if another alarm triggers a scale out policy during the cooldown period after a scale-in, application auto scaling scales out your scalable target immediately.

disableScaleIn :: Maybe Bool

Indicates whether scale in by the target tracking policy is disabled. If the value is true, scale in is disabled and the target tracking policy won't remove capacity from the scalable resource. Otherwise, scale in is enabled and the target tracking policy can remove capacity from the scalable resource. The default value is false.

scaleOutCooldown :: Maybe Int

The amount of time, in seconds, after a scale out activity completes before another scale out activity can start. While the cooldown period is in effect, the capacity that has been added by the previous scale out event that initiated the cooldown is calculated as part of the desired capacity for the next scale out. You should continuously (but not excessively) scale out.

targetValue :: Double

The target value for the metric. The range is 8.515920e-109 to 1.174271e+108 (Base 10) or 2e-360 to 2e360 (Base 2).

newAutoScalingTargetTrackingScalingPolicyConfigurationUpdate

Arguments

:: Double
$sel:targetValue:AutoScalingTargetTrackingScalingPolicyConfigurationUpdate'

-> AutoScalingTargetTrackingScalingPolicyConfigurationUpdate

Create a value of AutoScalingTargetTrackingScalingPolicyConfigurationUpdate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:scaleInCooldown:AutoScalingTargetTrackingScalingPolicyConfigurationUpdate', autoScalingTargetTrackingScalingPolicyConfigurationUpdate_scaleInCooldown - The amount of time, in seconds, after a scale in activity completes before another scale in activity can start. The cooldown period is used to block subsequent scale in requests until it has expired. You should scale in conservatively to protect your application's availability. However, if another alarm triggers a scale out policy during the cooldown period after a scale-in, application auto scaling scales out your scalable target immediately.

$sel:disableScaleIn:AutoScalingTargetTrackingScalingPolicyConfigurationUpdate', autoScalingTargetTrackingScalingPolicyConfigurationUpdate_disableScaleIn - Indicates whether scale in by the target tracking policy is disabled. If the value is true, scale in is disabled and the target tracking policy won't remove capacity from the scalable resource. Otherwise, scale in is enabled and the target tracking policy can remove capacity from the scalable resource. The default value is false.

$sel:scaleOutCooldown:AutoScalingTargetTrackingScalingPolicyConfigurationUpdate', autoScalingTargetTrackingScalingPolicyConfigurationUpdate_scaleOutCooldown - The amount of time, in seconds, after a scale out activity completes before another scale out activity can start. While the cooldown period is in effect, the capacity that has been added by the previous scale out event that initiated the cooldown is calculated as part of the desired capacity for the next scale out. You should continuously (but not excessively) scale out.

$sel:targetValue:AutoScalingTargetTrackingScalingPolicyConfigurationUpdate', autoScalingTargetTrackingScalingPolicyConfigurationUpdate_targetValue - The target value for the metric. The range is 8.515920e-109 to 1.174271e+108 (Base 10) or 2e-360 to 2e360 (Base 2).

autoScalingTargetTrackingScalingPolicyConfigurationUpdate_scaleInCooldown :: Lens' AutoScalingTargetTrackingScalingPolicyConfigurationUpdate (Maybe Int)

The amount of time, in seconds, after a scale in activity completes before another scale in activity can start. The cooldown period is used to block subsequent scale in requests until it has expired. You should scale in conservatively to protect your application's availability. However, if another alarm triggers a scale out policy during the cooldown period after a scale-in, application auto scaling scales out your scalable target immediately.

autoScalingTargetTrackingScalingPolicyConfigurationUpdate_disableScaleIn :: Lens' AutoScalingTargetTrackingScalingPolicyConfigurationUpdate (Maybe Bool)

Indicates whether scale in by the target tracking policy is disabled. If the value is true, scale in is disabled and the target tracking policy won't remove capacity from the scalable resource. Otherwise, scale in is enabled and the target tracking policy can remove capacity from the scalable resource. The default value is false.

autoScalingTargetTrackingScalingPolicyConfigurationUpdate_scaleOutCooldown :: Lens' AutoScalingTargetTrackingScalingPolicyConfigurationUpdate (Maybe Int)

The amount of time, in seconds, after a scale out activity completes before another scale out activity can start. While the cooldown period is in effect, the capacity that has been added by the previous scale out event that initiated the cooldown is calculated as part of the desired capacity for the next scale out. You should continuously (but not excessively) scale out.

autoScalingTargetTrackingScalingPolicyConfigurationUpdate_targetValue :: Lens' AutoScalingTargetTrackingScalingPolicyConfigurationUpdate Double

The target value for the metric. The range is 8.515920e-109 to 1.174271e+108 (Base 10) or 2e-360 to 2e360 (Base 2).