Network.AWS.DynamoDB.Types.BackupDescription

Description

Documentation

data BackupDescription

Contains the description of the backup created for the table.

See: newBackupDescription smart constructor.

Constructors

BackupDescription'

Fields

backupDetails :: Maybe BackupDetails

Contains the details of the backup created for the table.

sourceTableDetails :: Maybe SourceTableDetails

Contains the details of the table when the backup was created.

sourceTableFeatureDetails :: Maybe SourceTableFeatureDetails

Contains the details of the features enabled on the table when the backup was created. For example, LSIs, GSIs, streams, TTL.

newBackupDescription :: BackupDescription

Create a value of BackupDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:backupDetails:BackupDescription', backupDescription_backupDetails - Contains the details of the backup created for the table.

$sel:sourceTableDetails:BackupDescription', backupDescription_sourceTableDetails - Contains the details of the table when the backup was created.

$sel:sourceTableFeatureDetails:BackupDescription', backupDescription_sourceTableFeatureDetails - Contains the details of the features enabled on the table when the backup was created. For example, LSIs, GSIs, streams, TTL.

backupDescription_backupDetails :: Lens' BackupDescription (Maybe BackupDetails)

Contains the details of the backup created for the table.

backupDescription_sourceTableDetails :: Lens' BackupDescription (Maybe SourceTableDetails)

Contains the details of the table when the backup was created.

backupDescription_sourceTableFeatureDetails :: Lens' BackupDescription (Maybe SourceTableFeatureDetails)

Contains the details of the features enabled on the table when the backup was created. For example, LSIs, GSIs, streams, TTL.