Network.AWS.DynamoDB.Types.BackupSummary

Description

Documentation

data BackupSummary

Contains details for the backup.

See: newBackupSummary smart constructor.

Constructors

BackupSummary'

Fields

tableName :: Maybe Text

Name of the table.

backupName :: Maybe Text

Name of the specified backup.

backupCreationDateTime :: Maybe POSIX

Time at which the backup was created.

backupType :: Maybe BackupType

BackupType:

USER - You create and manage these using the on-demand backup feature.

SYSTEM - If you delete a table with point-in-time recovery enabled, a SYSTEM backup is automatically created and is retained for 35 days (at no additional cost). System backups allow you to restore the deleted table to the state it was in just before the point of deletion.

AWS_BACKUP - On-demand backup created by you from AWS Backup service.

tableId :: Maybe Text

Unique identifier for the table.

backupArn :: Maybe Text

ARN associated with the backup.

tableArn :: Maybe Text

ARN associated with the table.

backupExpiryDateTime :: Maybe POSIX

Time at which the automatic on-demand backup created by DynamoDB will expire. This SYSTEM on-demand backup expires automatically 35 days after its creation.

backupSizeBytes :: Maybe Natural

Size of the backup in bytes.

backupStatus :: Maybe BackupStatus

Backup can be in one of the following states: CREATING, ACTIVE, DELETED.

newBackupSummary :: BackupSummary

Create a value of BackupSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tableName:BackupSummary', backupSummary_tableName - Name of the table.

$sel:backupName:BackupSummary', backupSummary_backupName - Name of the specified backup.

$sel:backupCreationDateTime:BackupSummary', backupSummary_backupCreationDateTime - Time at which the backup was created.

$sel:backupType:BackupSummary', backupSummary_backupType - BackupType:

USER - You create and manage these using the on-demand backup feature.

SYSTEM - If you delete a table with point-in-time recovery enabled, a SYSTEM backup is automatically created and is retained for 35 days (at no additional cost). System backups allow you to restore the deleted table to the state it was in just before the point of deletion.

AWS_BACKUP - On-demand backup created by you from AWS Backup service.

$sel:tableId:BackupSummary', backupSummary_tableId - Unique identifier for the table.

$sel:backupArn:BackupSummary', backupSummary_backupArn - ARN associated with the backup.

$sel:tableArn:BackupSummary', backupSummary_tableArn - ARN associated with the table.

$sel:backupExpiryDateTime:BackupSummary', backupSummary_backupExpiryDateTime - Time at which the automatic on-demand backup created by DynamoDB will expire. This SYSTEM on-demand backup expires automatically 35 days after its creation.

$sel:backupSizeBytes:BackupSummary', backupSummary_backupSizeBytes - Size of the backup in bytes.

$sel:backupStatus:BackupSummary', backupSummary_backupStatus - Backup can be in one of the following states: CREATING, ACTIVE, DELETED.

backupSummary_tableName :: Lens' BackupSummary (Maybe Text)

Name of the table.

backupSummary_backupName :: Lens' BackupSummary (Maybe Text)

Name of the specified backup.

backupSummary_backupCreationDateTime :: Lens' BackupSummary (Maybe UTCTime)

Time at which the backup was created.

backupSummary_backupType :: Lens' BackupSummary (Maybe BackupType)

BackupType:

USER - You create and manage these using the on-demand backup feature.

SYSTEM - If you delete a table with point-in-time recovery enabled, a SYSTEM backup is automatically created and is retained for 35 days (at no additional cost). System backups allow you to restore the deleted table to the state it was in just before the point of deletion.

AWS_BACKUP - On-demand backup created by you from AWS Backup service.

backupSummary_tableId :: Lens' BackupSummary (Maybe Text)

Unique identifier for the table.

backupSummary_backupArn :: Lens' BackupSummary (Maybe Text)

ARN associated with the backup.

backupSummary_tableArn :: Lens' BackupSummary (Maybe Text)

ARN associated with the table.

backupSummary_backupExpiryDateTime :: Lens' BackupSummary (Maybe UTCTime)

Time at which the automatic on-demand backup created by DynamoDB will expire. This SYSTEM on-demand backup expires automatically 35 days after its creation.

backupSummary_backupSizeBytes :: Lens' BackupSummary (Maybe Natural)

Size of the backup in bytes.

backupSummary_backupStatus :: Lens' BackupSummary (Maybe BackupStatus)

Backup can be in one of the following states: CREATING, ACTIVE, DELETED.