Network.AWS.DynamoDB.Types.BillingModeSummary

Description

Documentation

data BillingModeSummary

Contains the details for the read/write capacity mode.

See: newBillingModeSummary smart constructor.

Constructors

BillingModeSummary'

Fields

lastUpdateToPayPerRequestDateTime :: Maybe POSIX

Represents the time when PAY_PER_REQUEST was last set as the read/write capacity mode.

billingMode :: Maybe BillingMode

Controls how you are charged for read and write throughput and how you manage capacity. This setting can be changed later.

PROVISIONED - Sets the read/write capacity mode to PROVISIONED. We recommend using PROVISIONED for predictable workloads.

PAY_PER_REQUEST - Sets the read/write capacity mode to PAY_PER_REQUEST. We recommend using PAY_PER_REQUEST for unpredictable workloads.

newBillingModeSummary :: BillingModeSummary

Create a value of BillingModeSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastUpdateToPayPerRequestDateTime:BillingModeSummary', billingModeSummary_lastUpdateToPayPerRequestDateTime - Represents the time when PAY_PER_REQUEST was last set as the read/write capacity mode.

$sel:billingMode:BillingModeSummary', billingModeSummary_billingMode - Controls how you are charged for read and write throughput and how you manage capacity. This setting can be changed later.

PROVISIONED - Sets the read/write capacity mode to PROVISIONED. We recommend using PROVISIONED for predictable workloads.

PAY_PER_REQUEST - Sets the read/write capacity mode to PAY_PER_REQUEST. We recommend using PAY_PER_REQUEST for unpredictable workloads.

billingModeSummary_lastUpdateToPayPerRequestDateTime :: Lens' BillingModeSummary (Maybe UTCTime)

Represents the time when PAY_PER_REQUEST was last set as the read/write capacity mode.

billingModeSummary_billingMode :: Lens' BillingModeSummary (Maybe BillingMode)

Controls how you are charged for read and write throughput and how you manage capacity. This setting can be changed later.

PROVISIONED - Sets the read/write capacity mode to PROVISIONED. We recommend using PROVISIONED for predictable workloads.

PAY_PER_REQUEST - Sets the read/write capacity mode to PAY_PER_REQUEST. We recommend using PAY_PER_REQUEST for unpredictable workloads.