Network.AWS.DynamoDB.Types.ConsumedCapacity

Description

Documentation

data ConsumedCapacity

The capacity units consumed by an operation. The data returned includes the total provisioned throughput consumed, along with statistics for the table and any indexes involved in the operation. ConsumedCapacity is only returned if the request asked for it. For more information, see Provisioned Throughput in the Amazon DynamoDB Developer Guide.

See: newConsumedCapacity smart constructor.

Constructors

ConsumedCapacity'

Fields

localSecondaryIndexes :: Maybe (HashMap Text Capacity)

The amount of throughput consumed on each local index affected by the operation.

globalSecondaryIndexes :: Maybe (HashMap Text Capacity)

The amount of throughput consumed on each global index affected by the operation.

tableName :: Maybe Text

The name of the table that was affected by the operation.

writeCapacityUnits :: Maybe Double

The total number of write capacity units consumed by the operation.

capacityUnits :: Maybe Double

The total number of capacity units consumed by the operation.

table :: Maybe Capacity

The amount of throughput consumed on the table affected by the operation.

readCapacityUnits :: Maybe Double

The total number of read capacity units consumed by the operation.

newConsumedCapacity :: ConsumedCapacity

Create a value of ConsumedCapacity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:localSecondaryIndexes:ConsumedCapacity', consumedCapacity_localSecondaryIndexes - The amount of throughput consumed on each local index affected by the operation.

$sel:globalSecondaryIndexes:ConsumedCapacity', consumedCapacity_globalSecondaryIndexes - The amount of throughput consumed on each global index affected by the operation.

$sel:tableName:ConsumedCapacity', consumedCapacity_tableName - The name of the table that was affected by the operation.

$sel:writeCapacityUnits:ConsumedCapacity', consumedCapacity_writeCapacityUnits - The total number of write capacity units consumed by the operation.

$sel:capacityUnits:ConsumedCapacity', consumedCapacity_capacityUnits - The total number of capacity units consumed by the operation.

$sel:table:ConsumedCapacity', consumedCapacity_table - The amount of throughput consumed on the table affected by the operation.

$sel:readCapacityUnits:ConsumedCapacity', consumedCapacity_readCapacityUnits - The total number of read capacity units consumed by the operation.

consumedCapacity_localSecondaryIndexes :: Lens' ConsumedCapacity (Maybe (HashMap Text Capacity))

The amount of throughput consumed on each local index affected by the operation.

consumedCapacity_globalSecondaryIndexes :: Lens' ConsumedCapacity (Maybe (HashMap Text Capacity))

The amount of throughput consumed on each global index affected by the operation.

consumedCapacity_tableName :: Lens' ConsumedCapacity (Maybe Text)

The name of the table that was affected by the operation.

consumedCapacity_writeCapacityUnits :: Lens' ConsumedCapacity (Maybe Double)

The total number of write capacity units consumed by the operation.

consumedCapacity_capacityUnits :: Lens' ConsumedCapacity (Maybe Double)

The total number of capacity units consumed by the operation.

consumedCapacity_table :: Lens' ConsumedCapacity (Maybe Capacity)

The amount of throughput consumed on the table affected by the operation.

consumedCapacity_readCapacityUnits :: Lens' ConsumedCapacity (Maybe Double)

The total number of read capacity units consumed by the operation.