Network.AWS.DynamoDB.Types.CreateReplicationGroupMemberAction

Description

Documentation

data CreateReplicationGroupMemberAction

Represents a replica to be created.

See: newCreateReplicationGroupMemberAction smart constructor.

Constructors

CreateReplicationGroupMemberAction'

Fields

globalSecondaryIndexes :: Maybe (NonEmpty ReplicaGlobalSecondaryIndex)

Replica-specific global secondary index settings.

provisionedThroughputOverride :: Maybe ProvisionedThroughputOverride

Replica-specific provisioned throughput. If not specified, uses the source table's provisioned throughput settings.

kmsMasterKeyId :: Maybe Text

The AWS KMS customer master key (CMK) that should be used for AWS KMS encryption in the new replica. To specify a CMK, use its key ID, Amazon Resource Name (ARN), alias name, or alias ARN. Note that you should only provide this parameter if the key is different from the default DynamoDB KMS master key alias/aws/dynamodb.

regionName :: Text

The Region where the new replica will be created.

newCreateReplicationGroupMemberAction

Arguments

:: Text
$sel:regionName:CreateReplicationGroupMemberAction'

-> CreateReplicationGroupMemberAction

Create a value of CreateReplicationGroupMemberAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:globalSecondaryIndexes:CreateReplicationGroupMemberAction', createReplicationGroupMemberAction_globalSecondaryIndexes - Replica-specific global secondary index settings.

$sel:provisionedThroughputOverride:CreateReplicationGroupMemberAction', createReplicationGroupMemberAction_provisionedThroughputOverride - Replica-specific provisioned throughput. If not specified, uses the source table's provisioned throughput settings.

$sel:kmsMasterKeyId:CreateReplicationGroupMemberAction', createReplicationGroupMemberAction_kmsMasterKeyId - The AWS KMS customer master key (CMK) that should be used for AWS KMS encryption in the new replica. To specify a CMK, use its key ID, Amazon Resource Name (ARN), alias name, or alias ARN. Note that you should only provide this parameter if the key is different from the default DynamoDB KMS master key alias/aws/dynamodb.

$sel:regionName:CreateReplicationGroupMemberAction', createReplicationGroupMemberAction_regionName - The Region where the new replica will be created.

createReplicationGroupMemberAction_globalSecondaryIndexes :: Lens' CreateReplicationGroupMemberAction (Maybe (NonEmpty ReplicaGlobalSecondaryIndex))

Replica-specific global secondary index settings.

createReplicationGroupMemberAction_provisionedThroughputOverride :: Lens' CreateReplicationGroupMemberAction (Maybe ProvisionedThroughputOverride)

Replica-specific provisioned throughput. If not specified, uses the source table's provisioned throughput settings.

createReplicationGroupMemberAction_kmsMasterKeyId :: Lens' CreateReplicationGroupMemberAction (Maybe Text)

The AWS KMS customer master key (CMK) that should be used for AWS KMS encryption in the new replica. To specify a CMK, use its key ID, Amazon Resource Name (ARN), alias name, or alias ARN. Note that you should only provide this parameter if the key is different from the default DynamoDB KMS master key alias/aws/dynamodb.

createReplicationGroupMemberAction_regionName :: Lens' CreateReplicationGroupMemberAction Text

The Region where the new replica will be created.