Network.AWS.DynamoDB.Types.Get

Description

Documentation

data Get

Specifies an item and related attribute values to retrieve in a TransactGetItem object.

See: newGet smart constructor.

Constructors

Get'

Fields

projectionExpression :: Maybe Text

A string that identifies one or more attributes of the specified item to retrieve from the table. The attributes in the expression must be separated by commas. If no attribute names are specified, then all attributes of the specified item are returned. If any of the requested attributes are not found, they do not appear in the result.

expressionAttributeNames :: Maybe (HashMap Text Text)

One or more substitution tokens for attribute names in the ProjectionExpression parameter.

key :: HashMap Text AttributeValue

A map of attribute names to AttributeValue objects that specifies the primary key of the item to retrieve.

tableName :: Text

The name of the table from which to retrieve the specified item.

newGet

Arguments

:: Text
$sel:tableName:Get'

-> Get

Create a value of Get with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:projectionExpression:Get', get_projectionExpression - A string that identifies one or more attributes of the specified item to retrieve from the table. The attributes in the expression must be separated by commas. If no attribute names are specified, then all attributes of the specified item are returned. If any of the requested attributes are not found, they do not appear in the result.

$sel:expressionAttributeNames:Get', get_expressionAttributeNames - One or more substitution tokens for attribute names in the ProjectionExpression parameter.

$sel:key:Get', get_key - A map of attribute names to AttributeValue objects that specifies the primary key of the item to retrieve.

$sel:tableName:Get', get_tableName - The name of the table from which to retrieve the specified item.

get_projectionExpression :: Lens' Get (Maybe Text)

A string that identifies one or more attributes of the specified item to retrieve from the table. The attributes in the expression must be separated by commas. If no attribute names are specified, then all attributes of the specified item are returned. If any of the requested attributes are not found, they do not appear in the result.

get_expressionAttributeNames :: Lens' Get (Maybe (HashMap Text Text))

One or more substitution tokens for attribute names in the ProjectionExpression parameter.

get_key :: Lens' Get (HashMap Text AttributeValue)

A map of attribute names to AttributeValue objects that specifies the primary key of the item to retrieve.

get_tableName :: Lens' Get Text

The name of the table from which to retrieve the specified item.