Network.AWS.DynamoDB.Types.GlobalSecondaryIndexDescription

Description

Documentation

data GlobalSecondaryIndexDescription

Represents the properties of a global secondary index.

See: newGlobalSecondaryIndexDescription smart constructor.

Constructors

GlobalSecondaryIndexDescription'

Fields

indexName :: Maybe Text

The name of the global secondary index.

keySchema :: Maybe (NonEmpty KeySchemaElement)

The complete key schema for a global secondary index, which consists of one or more pairs of attribute names and key types:

HASH - partition key

RANGE - sort key

The partition key of an item is also known as its hash attribute. The term "hash attribute" derives from DynamoDB's usage of an internal hash function to evenly distribute data items across partitions, based on their partition key values.

The sort key of an item is also known as its range attribute. The term "range attribute" derives from the way DynamoDB stores items with the same partition key physically close together, in sorted order by the sort key value.

indexArn :: Maybe Text

The Amazon Resource Name (ARN) that uniquely identifies the index.

projection :: Maybe Projection

Represents attributes that are copied (projected) from the table into the global secondary index. These are in addition to the primary key attributes and index key attributes, which are automatically projected.

indexSizeBytes :: Maybe Integer

The total size of the specified index, in bytes. DynamoDB updates this value approximately every six hours. Recent changes might not be reflected in this value.

backfilling :: Maybe Bool

Indicates whether the index is currently backfilling. Backfilling is the process of reading items from the table and determining whether they can be added to the index. (Not all items will qualify: For example, a partition key cannot have any duplicate values.) If an item can be added to the index, DynamoDB will do so. After all items have been processed, the backfilling operation is complete and Backfilling is false.

You can delete an index that is being created during the Backfilling phase when IndexStatus is set to CREATING and Backfilling is true. You can't delete the index that is being created when IndexStatus is set to CREATING and Backfilling is false.

For indexes that were created during a CreateTable operation, the Backfilling attribute does not appear in the DescribeTable output.

itemCount :: Maybe Integer

The number of items in the specified index. DynamoDB updates this value approximately every six hours. Recent changes might not be reflected in this value.

indexStatus :: Maybe IndexStatus

The current state of the global secondary index:

CREATING - The index is being created.

UPDATING - The index is being updated.

DELETING - The index is being deleted.

ACTIVE - The index is ready for use.

provisionedThroughput :: Maybe ProvisionedThroughputDescription

Represents the provisioned throughput settings for the specified global secondary index.

For current minimum and maximum provisioned throughput values, see Service, Account, and Table Quotas in the Amazon DynamoDB Developer Guide.

newGlobalSecondaryIndexDescription :: GlobalSecondaryIndexDescription

Create a value of GlobalSecondaryIndexDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:indexName:GlobalSecondaryIndexDescription', globalSecondaryIndexDescription_indexName - The name of the global secondary index.

$sel:keySchema:GlobalSecondaryIndexDescription', globalSecondaryIndexDescription_keySchema - The complete key schema for a global secondary index, which consists of one or more pairs of attribute names and key types:

HASH - partition key

RANGE - sort key

The partition key of an item is also known as its hash attribute. The term "hash attribute" derives from DynamoDB's usage of an internal hash function to evenly distribute data items across partitions, based on their partition key values.

The sort key of an item is also known as its range attribute. The term "range attribute" derives from the way DynamoDB stores items with the same partition key physically close together, in sorted order by the sort key value.

$sel:indexArn:GlobalSecondaryIndexDescription', globalSecondaryIndexDescription_indexArn - The Amazon Resource Name (ARN) that uniquely identifies the index.

$sel:projection:GlobalSecondaryIndexDescription', globalSecondaryIndexDescription_projection - Represents attributes that are copied (projected) from the table into the global secondary index. These are in addition to the primary key attributes and index key attributes, which are automatically projected.

$sel:indexSizeBytes:GlobalSecondaryIndexDescription', globalSecondaryIndexDescription_indexSizeBytes - The total size of the specified index, in bytes. DynamoDB updates this value approximately every six hours. Recent changes might not be reflected in this value.

$sel:backfilling:GlobalSecondaryIndexDescription', globalSecondaryIndexDescription_backfilling - Indicates whether the index is currently backfilling. Backfilling is the process of reading items from the table and determining whether they can be added to the index. (Not all items will qualify: For example, a partition key cannot have any duplicate values.) If an item can be added to the index, DynamoDB will do so. After all items have been processed, the backfilling operation is complete and Backfilling is false.

You can delete an index that is being created during the Backfilling phase when IndexStatus is set to CREATING and Backfilling is true. You can't delete the index that is being created when IndexStatus is set to CREATING and Backfilling is false.

For indexes that were created during a CreateTable operation, the Backfilling attribute does not appear in the DescribeTable output.

$sel:itemCount:GlobalSecondaryIndexDescription', globalSecondaryIndexDescription_itemCount - The number of items in the specified index. DynamoDB updates this value approximately every six hours. Recent changes might not be reflected in this value.

$sel:indexStatus:GlobalSecondaryIndexDescription', globalSecondaryIndexDescription_indexStatus - The current state of the global secondary index:

CREATING - The index is being created.

UPDATING - The index is being updated.

DELETING - The index is being deleted.

ACTIVE - The index is ready for use.

$sel:provisionedThroughput:GlobalSecondaryIndexDescription', globalSecondaryIndexDescription_provisionedThroughput - Represents the provisioned throughput settings for the specified global secondary index.

For current minimum and maximum provisioned throughput values, see Service, Account, and Table Quotas in the Amazon DynamoDB Developer Guide.

globalSecondaryIndexDescription_indexName :: Lens' GlobalSecondaryIndexDescription (Maybe Text)

The name of the global secondary index.

globalSecondaryIndexDescription_keySchema :: Lens' GlobalSecondaryIndexDescription (Maybe (NonEmpty KeySchemaElement))

The complete key schema for a global secondary index, which consists of one or more pairs of attribute names and key types:

HASH - partition key

RANGE - sort key

The partition key of an item is also known as its hash attribute. The term "hash attribute" derives from DynamoDB's usage of an internal hash function to evenly distribute data items across partitions, based on their partition key values.

The sort key of an item is also known as its range attribute. The term "range attribute" derives from the way DynamoDB stores items with the same partition key physically close together, in sorted order by the sort key value.

globalSecondaryIndexDescription_indexArn :: Lens' GlobalSecondaryIndexDescription (Maybe Text)

The Amazon Resource Name (ARN) that uniquely identifies the index.

globalSecondaryIndexDescription_projection :: Lens' GlobalSecondaryIndexDescription (Maybe Projection)

Represents attributes that are copied (projected) from the table into the global secondary index. These are in addition to the primary key attributes and index key attributes, which are automatically projected.

globalSecondaryIndexDescription_indexSizeBytes :: Lens' GlobalSecondaryIndexDescription (Maybe Integer)

The total size of the specified index, in bytes. DynamoDB updates this value approximately every six hours. Recent changes might not be reflected in this value.

globalSecondaryIndexDescription_backfilling :: Lens' GlobalSecondaryIndexDescription (Maybe Bool)

Indicates whether the index is currently backfilling. Backfilling is the process of reading items from the table and determining whether they can be added to the index. (Not all items will qualify: For example, a partition key cannot have any duplicate values.) If an item can be added to the index, DynamoDB will do so. After all items have been processed, the backfilling operation is complete and Backfilling is false.

You can delete an index that is being created during the Backfilling phase when IndexStatus is set to CREATING and Backfilling is true. You can't delete the index that is being created when IndexStatus is set to CREATING and Backfilling is false.

For indexes that were created during a CreateTable operation, the Backfilling attribute does not appear in the DescribeTable output.

globalSecondaryIndexDescription_itemCount :: Lens' GlobalSecondaryIndexDescription (Maybe Integer)

The number of items in the specified index. DynamoDB updates this value approximately every six hours. Recent changes might not be reflected in this value.

globalSecondaryIndexDescription_indexStatus :: Lens' GlobalSecondaryIndexDescription (Maybe IndexStatus)

The current state of the global secondary index:

CREATING - The index is being created.

UPDATING - The index is being updated.

DELETING - The index is being deleted.

ACTIVE - The index is ready for use.

globalSecondaryIndexDescription_provisionedThroughput :: Lens' GlobalSecondaryIndexDescription (Maybe ProvisionedThroughputDescription)

Represents the provisioned throughput settings for the specified global secondary index.

For current minimum and maximum provisioned throughput values, see Service, Account, and Table Quotas in the Amazon DynamoDB Developer Guide.