Network.AWS.DynamoDB.Types.ItemCollectionMetrics

Description

Documentation

data ItemCollectionMetrics

Information about item collections, if any, that were affected by the operation. ItemCollectionMetrics is only returned if the request asked for it. If the table does not have any local secondary indexes, this information is not returned in the response.

See: newItemCollectionMetrics smart constructor.

Constructors

ItemCollectionMetrics'

Fields

itemCollectionKey :: Maybe (HashMap Text AttributeValue)

The partition key value of the item collection. This value is the same as the partition key value of the item.

sizeEstimateRangeGB :: Maybe [Double]

An estimate of item collection size, in gigabytes. This value is a two-element array containing a lower bound and an upper bound for the estimate. The estimate includes the size of all the items in the table, plus the size of all attributes projected into all of the local secondary indexes on that table. Use this estimate to measure whether a local secondary index is approaching its size limit.

The estimate is subject to change over time; therefore, do not rely on the precision or accuracy of the estimate.

newItemCollectionMetrics :: ItemCollectionMetrics

Create a value of ItemCollectionMetrics with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:itemCollectionKey:ItemCollectionMetrics', itemCollectionMetrics_itemCollectionKey - The partition key value of the item collection. This value is the same as the partition key value of the item.

$sel:sizeEstimateRangeGB:ItemCollectionMetrics', itemCollectionMetrics_sizeEstimateRangeGB - An estimate of item collection size, in gigabytes. This value is a two-element array containing a lower bound and an upper bound for the estimate. The estimate includes the size of all the items in the table, plus the size of all attributes projected into all of the local secondary indexes on that table. Use this estimate to measure whether a local secondary index is approaching its size limit.

The estimate is subject to change over time; therefore, do not rely on the precision or accuracy of the estimate.

itemCollectionMetrics_itemCollectionKey :: Lens' ItemCollectionMetrics (Maybe (HashMap Text AttributeValue))

The partition key value of the item collection. This value is the same as the partition key value of the item.

itemCollectionMetrics_sizeEstimateRangeGB :: Lens' ItemCollectionMetrics (Maybe [Double])

An estimate of item collection size, in gigabytes. This value is a two-element array containing a lower bound and an upper bound for the estimate. The estimate includes the size of all the items in the table, plus the size of all attributes projected into all of the local secondary indexes on that table. Use this estimate to measure whether a local secondary index is approaching its size limit.

The estimate is subject to change over time; therefore, do not rely on the precision or accuracy of the estimate.