Network.AWS.DynamoDB.Types.KeysAndAttributes

Description

Documentation

data KeysAndAttributes

Represents a set of primary keys and, for each key, the attributes to retrieve from the table.

For each primary key, you must provide all of the key attributes. For example, with a simple primary key, you only need to provide the partition key. For a composite primary key, you must provide both the partition key and the sort key.

See: newKeysAndAttributes smart constructor.

Constructors

KeysAndAttributes'

Fields

projectionExpression :: Maybe Text

A string that identifies one or more attributes to retrieve from the table. These attributes can include scalars, sets, or elements of a JSON document. The attributes in the ProjectionExpression must be separated by commas.

If no attribute names are specified, then all attributes will be returned. If any of the requested attributes are not found, they will not appear in the result.

For more information, see Accessing Item Attributes in the Amazon DynamoDB Developer Guide.

expressionAttributeNames :: Maybe (HashMap Text Text)

One or more substitution tokens for attribute names in an expression. The following are some use cases for using ExpressionAttributeNames:

To access an attribute whose name conflicts with a DynamoDB reserved word.

To create a placeholder for repeating occurrences of an attribute name in an expression.

To prevent special characters in an attribute name from being misinterpreted in an expression.

Use the # character in an expression to dereference an attribute name. For example, consider the following attribute name:

Percentile

The name of this attribute conflicts with a reserved word, so it cannot be used directly in an expression. (For the complete list of reserved words, see Reserved Words in the Amazon DynamoDB Developer Guide). To work around this, you could specify the following for ExpressionAttributeNames:

{"#P":"Percentile"}

You could then use this substitution in an expression, as in this example:

#P = :val

Tokens that begin with the : character are /expression attribute values/, which are placeholders for the actual value at runtime.

For more information on expression attribute names, see Accessing Item Attributes in the Amazon DynamoDB Developer Guide.

consistentRead :: Maybe Bool

The consistency of a read operation. If set to true, then a strongly consistent read is used; otherwise, an eventually consistent read is used.

attributesToGet :: Maybe (NonEmpty Text)

This is a legacy parameter. Use ProjectionExpression instead. For more information, see Legacy Conditional Parameters in the Amazon DynamoDB Developer Guide.

keys :: NonEmpty (HashMap Text AttributeValue)

The primary key attribute values that define the items and the attributes associated with the items.

newKeysAndAttributes

Arguments

:: NonEmpty (HashMap Text AttributeValue)
$sel:keys:KeysAndAttributes'

-> KeysAndAttributes

Create a value of KeysAndAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:projectionExpression:KeysAndAttributes', keysAndAttributes_projectionExpression - A string that identifies one or more attributes to retrieve from the table. These attributes can include scalars, sets, or elements of a JSON document. The attributes in the ProjectionExpression must be separated by commas.

If no attribute names are specified, then all attributes will be returned. If any of the requested attributes are not found, they will not appear in the result.

For more information, see Accessing Item Attributes in the Amazon DynamoDB Developer Guide.

$sel:expressionAttributeNames:KeysAndAttributes', keysAndAttributes_expressionAttributeNames - One or more substitution tokens for attribute names in an expression. The following are some use cases for using ExpressionAttributeNames:

To access an attribute whose name conflicts with a DynamoDB reserved word.

To create a placeholder for repeating occurrences of an attribute name in an expression.

To prevent special characters in an attribute name from being misinterpreted in an expression.

Use the # character in an expression to dereference an attribute name. For example, consider the following attribute name:

Percentile

The name of this attribute conflicts with a reserved word, so it cannot be used directly in an expression. (For the complete list of reserved words, see Reserved Words in the Amazon DynamoDB Developer Guide). To work around this, you could specify the following for ExpressionAttributeNames:

{"#P":"Percentile"}

You could then use this substitution in an expression, as in this example:

#P = :val

Tokens that begin with the : character are /expression attribute values/, which are placeholders for the actual value at runtime.

For more information on expression attribute names, see Accessing Item Attributes in the Amazon DynamoDB Developer Guide.

$sel:consistentRead:KeysAndAttributes', keysAndAttributes_consistentRead - The consistency of a read operation. If set to true, then a strongly consistent read is used; otherwise, an eventually consistent read is used.

$sel:attributesToGet:KeysAndAttributes', keysAndAttributes_attributesToGet - This is a legacy parameter. Use ProjectionExpression instead. For more information, see Legacy Conditional Parameters in the Amazon DynamoDB Developer Guide.

$sel:keys:KeysAndAttributes', keysAndAttributes_keys - The primary key attribute values that define the items and the attributes associated with the items.

keysAndAttributes_projectionExpression :: Lens' KeysAndAttributes (Maybe Text)

A string that identifies one or more attributes to retrieve from the table. These attributes can include scalars, sets, or elements of a JSON document. The attributes in the ProjectionExpression must be separated by commas.

If no attribute names are specified, then all attributes will be returned. If any of the requested attributes are not found, they will not appear in the result.

For more information, see Accessing Item Attributes in the Amazon DynamoDB Developer Guide.

keysAndAttributes_expressionAttributeNames :: Lens' KeysAndAttributes (Maybe (HashMap Text Text))

One or more substitution tokens for attribute names in an expression. The following are some use cases for using ExpressionAttributeNames:

To access an attribute whose name conflicts with a DynamoDB reserved word.

To create a placeholder for repeating occurrences of an attribute name in an expression.

To prevent special characters in an attribute name from being misinterpreted in an expression.

Use the # character in an expression to dereference an attribute name. For example, consider the following attribute name:

Percentile

The name of this attribute conflicts with a reserved word, so it cannot be used directly in an expression. (For the complete list of reserved words, see Reserved Words in the Amazon DynamoDB Developer Guide). To work around this, you could specify the following for ExpressionAttributeNames:

{"#P":"Percentile"}

You could then use this substitution in an expression, as in this example:

#P = :val

Tokens that begin with the : character are /expression attribute values/, which are placeholders for the actual value at runtime.

For more information on expression attribute names, see Accessing Item Attributes in the Amazon DynamoDB Developer Guide.

keysAndAttributes_consistentRead :: Lens' KeysAndAttributes (Maybe Bool)

The consistency of a read operation. If set to true, then a strongly consistent read is used; otherwise, an eventually consistent read is used.

keysAndAttributes_attributesToGet :: Lens' KeysAndAttributes (Maybe (NonEmpty Text))

This is a legacy parameter. Use ProjectionExpression instead. For more information, see Legacy Conditional Parameters in the Amazon DynamoDB Developer Guide.

keysAndAttributes_keys :: Lens' KeysAndAttributes (NonEmpty (HashMap Text AttributeValue))

The primary key attribute values that define the items and the attributes associated with the items.