Network.AWS.DynamoDB.Types.LocalSecondaryIndex

Description

Documentation

data LocalSecondaryIndex

Represents the properties of a local secondary index.

See: newLocalSecondaryIndex smart constructor.

Constructors

LocalSecondaryIndex'

Fields

indexName :: Text

The name of the local secondary index. The name must be unique among all other indexes on this table.

keySchema :: NonEmpty KeySchemaElement

The complete key schema for the local secondary index, consisting of one or more pairs of attribute names and key types:

HASH - partition key

RANGE - sort key

The partition key of an item is also known as its hash attribute. The term "hash attribute" derives from DynamoDB's usage of an internal hash function to evenly distribute data items across partitions, based on their partition key values.

The sort key of an item is also known as its range attribute. The term "range attribute" derives from the way DynamoDB stores items with the same partition key physically close together, in sorted order by the sort key value.

projection :: Projection

Represents attributes that are copied (projected) from the table into the local secondary index. These are in addition to the primary key attributes and index key attributes, which are automatically projected.

newLocalSecondaryIndex

Arguments

:: Text
$sel:indexName:LocalSecondaryIndex'

-> NonEmpty KeySchemaElement
$sel:keySchema:LocalSecondaryIndex'

-> Projection
$sel:projection:LocalSecondaryIndex'

-> LocalSecondaryIndex

Create a value of LocalSecondaryIndex with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:indexName:LocalSecondaryIndex', localSecondaryIndex_indexName - The name of the local secondary index. The name must be unique among all other indexes on this table.

$sel:keySchema:LocalSecondaryIndex', localSecondaryIndex_keySchema - The complete key schema for the local secondary index, consisting of one or more pairs of attribute names and key types:

HASH - partition key

RANGE - sort key

The partition key of an item is also known as its hash attribute. The term "hash attribute" derives from DynamoDB's usage of an internal hash function to evenly distribute data items across partitions, based on their partition key values.

The sort key of an item is also known as its range attribute. The term "range attribute" derives from the way DynamoDB stores items with the same partition key physically close together, in sorted order by the sort key value.

$sel:projection:LocalSecondaryIndex', localSecondaryIndex_projection - Represents attributes that are copied (projected) from the table into the local secondary index. These are in addition to the primary key attributes and index key attributes, which are automatically projected.

localSecondaryIndex_indexName :: Lens' LocalSecondaryIndex Text

The name of the local secondary index. The name must be unique among all other indexes on this table.

localSecondaryIndex_keySchema :: Lens' LocalSecondaryIndex (NonEmpty KeySchemaElement)

The complete key schema for the local secondary index, consisting of one or more pairs of attribute names and key types:

HASH - partition key

RANGE - sort key

The partition key of an item is also known as its hash attribute. The term "hash attribute" derives from DynamoDB's usage of an internal hash function to evenly distribute data items across partitions, based on their partition key values.

The sort key of an item is also known as its range attribute. The term "range attribute" derives from the way DynamoDB stores items with the same partition key physically close together, in sorted order by the sort key value.

localSecondaryIndex_projection :: Lens' LocalSecondaryIndex Projection

Represents attributes that are copied (projected) from the table into the local secondary index. These are in addition to the primary key attributes and index key attributes, which are automatically projected.