Network.AWS.DynamoDB.Types.LocalSecondaryIndexDescription

Description

Documentation

data LocalSecondaryIndexDescription

Represents the properties of a local secondary index.

See: newLocalSecondaryIndexDescription smart constructor.

Constructors

LocalSecondaryIndexDescription'

Fields

indexName :: Maybe Text

Represents the name of the local secondary index.

keySchema :: Maybe (NonEmpty KeySchemaElement)

The complete key schema for the local secondary index, consisting of one or more pairs of attribute names and key types:

HASH - partition key

RANGE - sort key

The partition key of an item is also known as its hash attribute. The term "hash attribute" derives from DynamoDB's usage of an internal hash function to evenly distribute data items across partitions, based on their partition key values.

The sort key of an item is also known as its range attribute. The term "range attribute" derives from the way DynamoDB stores items with the same partition key physically close together, in sorted order by the sort key value.

indexArn :: Maybe Text

The Amazon Resource Name (ARN) that uniquely identifies the index.

projection :: Maybe Projection

Represents attributes that are copied (projected) from the table into the global secondary index. These are in addition to the primary key attributes and index key attributes, which are automatically projected.

indexSizeBytes :: Maybe Integer

The total size of the specified index, in bytes. DynamoDB updates this value approximately every six hours. Recent changes might not be reflected in this value.

itemCount :: Maybe Integer

The number of items in the specified index. DynamoDB updates this value approximately every six hours. Recent changes might not be reflected in this value.

newLocalSecondaryIndexDescription :: LocalSecondaryIndexDescription

Create a value of LocalSecondaryIndexDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:indexName:LocalSecondaryIndexDescription', localSecondaryIndexDescription_indexName - Represents the name of the local secondary index.

$sel:keySchema:LocalSecondaryIndexDescription', localSecondaryIndexDescription_keySchema - The complete key schema for the local secondary index, consisting of one or more pairs of attribute names and key types:

HASH - partition key

RANGE - sort key

The partition key of an item is also known as its hash attribute. The term "hash attribute" derives from DynamoDB's usage of an internal hash function to evenly distribute data items across partitions, based on their partition key values.

The sort key of an item is also known as its range attribute. The term "range attribute" derives from the way DynamoDB stores items with the same partition key physically close together, in sorted order by the sort key value.

$sel:indexArn:LocalSecondaryIndexDescription', localSecondaryIndexDescription_indexArn - The Amazon Resource Name (ARN) that uniquely identifies the index.

$sel:projection:LocalSecondaryIndexDescription', localSecondaryIndexDescription_projection - Represents attributes that are copied (projected) from the table into the global secondary index. These are in addition to the primary key attributes and index key attributes, which are automatically projected.

$sel:indexSizeBytes:LocalSecondaryIndexDescription', localSecondaryIndexDescription_indexSizeBytes - The total size of the specified index, in bytes. DynamoDB updates this value approximately every six hours. Recent changes might not be reflected in this value.

$sel:itemCount:LocalSecondaryIndexDescription', localSecondaryIndexDescription_itemCount - The number of items in the specified index. DynamoDB updates this value approximately every six hours. Recent changes might not be reflected in this value.

localSecondaryIndexDescription_indexName :: Lens' LocalSecondaryIndexDescription (Maybe Text)

Represents the name of the local secondary index.

localSecondaryIndexDescription_keySchema :: Lens' LocalSecondaryIndexDescription (Maybe (NonEmpty KeySchemaElement))

The complete key schema for the local secondary index, consisting of one or more pairs of attribute names and key types:

HASH - partition key

RANGE - sort key

The partition key of an item is also known as its hash attribute. The term "hash attribute" derives from DynamoDB's usage of an internal hash function to evenly distribute data items across partitions, based on their partition key values.

The sort key of an item is also known as its range attribute. The term "range attribute" derives from the way DynamoDB stores items with the same partition key physically close together, in sorted order by the sort key value.

localSecondaryIndexDescription_indexArn :: Lens' LocalSecondaryIndexDescription (Maybe Text)

The Amazon Resource Name (ARN) that uniquely identifies the index.

localSecondaryIndexDescription_projection :: Lens' LocalSecondaryIndexDescription (Maybe Projection)

Represents attributes that are copied (projected) from the table into the global secondary index. These are in addition to the primary key attributes and index key attributes, which are automatically projected.

localSecondaryIndexDescription_indexSizeBytes :: Lens' LocalSecondaryIndexDescription (Maybe Integer)

The total size of the specified index, in bytes. DynamoDB updates this value approximately every six hours. Recent changes might not be reflected in this value.

localSecondaryIndexDescription_itemCount :: Lens' LocalSecondaryIndexDescription (Maybe Integer)

The number of items in the specified index. DynamoDB updates this value approximately every six hours. Recent changes might not be reflected in this value.