Network.AWS.DynamoDB.Types.ProvisionedThroughputDescription

Description

Documentation

data ProvisionedThroughputDescription

Represents the provisioned throughput settings for the table, consisting of read and write capacity units, along with data about increases and decreases.

See: newProvisionedThroughputDescription smart constructor.

Constructors

ProvisionedThroughputDescription'

Fields

lastDecreaseDateTime :: Maybe POSIX

The date and time of the last provisioned throughput decrease for this table.

lastIncreaseDateTime :: Maybe POSIX

The date and time of the last provisioned throughput increase for this table.

writeCapacityUnits :: Maybe Natural

The maximum number of writes consumed per second before DynamoDB returns a ThrottlingException.

numberOfDecreasesToday :: Maybe Natural

The number of provisioned throughput decreases for this table during this UTC calendar day. For current maximums on provisioned throughput decreases, see Service, Account, and Table Quotas in the Amazon DynamoDB Developer Guide.

readCapacityUnits :: Maybe Natural

The maximum number of strongly consistent reads consumed per second before DynamoDB returns a ThrottlingException. Eventually consistent reads require less effort than strongly consistent reads, so a setting of 50 ReadCapacityUnits per second provides 100 eventually consistent ReadCapacityUnits per second.

newProvisionedThroughputDescription :: ProvisionedThroughputDescription

Create a value of ProvisionedThroughputDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastDecreaseDateTime:ProvisionedThroughputDescription', provisionedThroughputDescription_lastDecreaseDateTime - The date and time of the last provisioned throughput decrease for this table.

$sel:lastIncreaseDateTime:ProvisionedThroughputDescription', provisionedThroughputDescription_lastIncreaseDateTime - The date and time of the last provisioned throughput increase for this table.

$sel:writeCapacityUnits:ProvisionedThroughputDescription', provisionedThroughputDescription_writeCapacityUnits - The maximum number of writes consumed per second before DynamoDB returns a ThrottlingException.

$sel:numberOfDecreasesToday:ProvisionedThroughputDescription', provisionedThroughputDescription_numberOfDecreasesToday - The number of provisioned throughput decreases for this table during this UTC calendar day. For current maximums on provisioned throughput decreases, see Service, Account, and Table Quotas in the Amazon DynamoDB Developer Guide.

$sel:readCapacityUnits:ProvisionedThroughputDescription', provisionedThroughputDescription_readCapacityUnits - The maximum number of strongly consistent reads consumed per second before DynamoDB returns a ThrottlingException. Eventually consistent reads require less effort than strongly consistent reads, so a setting of 50 ReadCapacityUnits per second provides 100 eventually consistent ReadCapacityUnits per second.

provisionedThroughputDescription_lastDecreaseDateTime :: Lens' ProvisionedThroughputDescription (Maybe UTCTime)

The date and time of the last provisioned throughput decrease for this table.

provisionedThroughputDescription_lastIncreaseDateTime :: Lens' ProvisionedThroughputDescription (Maybe UTCTime)

The date and time of the last provisioned throughput increase for this table.

provisionedThroughputDescription_writeCapacityUnits :: Lens' ProvisionedThroughputDescription (Maybe Natural)

The maximum number of writes consumed per second before DynamoDB returns a ThrottlingException.

provisionedThroughputDescription_numberOfDecreasesToday :: Lens' ProvisionedThroughputDescription (Maybe Natural)

The number of provisioned throughput decreases for this table during this UTC calendar day. For current maximums on provisioned throughput decreases, see Service, Account, and Table Quotas in the Amazon DynamoDB Developer Guide.

provisionedThroughputDescription_readCapacityUnits :: Lens' ProvisionedThroughputDescription (Maybe Natural)

The maximum number of strongly consistent reads consumed per second before DynamoDB returns a ThrottlingException. Eventually consistent reads require less effort than strongly consistent reads, so a setting of 50 ReadCapacityUnits per second provides 100 eventually consistent ReadCapacityUnits per second.