Network.AWS.DynamoDB.Types.ReplicaAutoScalingUpdate

Description

Documentation

data ReplicaAutoScalingUpdate

Represents the auto scaling settings of a replica that will be modified.

See: newReplicaAutoScalingUpdate smart constructor.

Constructors

ReplicaAutoScalingUpdate'

Fields

replicaProvisionedReadCapacityAutoScalingUpdate :: Maybe AutoScalingSettingsUpdate

replicaGlobalSecondaryIndexUpdates :: Maybe [ReplicaGlobalSecondaryIndexAutoScalingUpdate]

Represents the auto scaling settings of global secondary indexes that will be modified.

regionName :: Text

The Region where the replica exists.

newReplicaAutoScalingUpdate

Arguments

:: Text
$sel:regionName:ReplicaAutoScalingUpdate'

-> ReplicaAutoScalingUpdate

Create a value of ReplicaAutoScalingUpdate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:replicaProvisionedReadCapacityAutoScalingUpdate:ReplicaAutoScalingUpdate', replicaAutoScalingUpdate_replicaProvisionedReadCapacityAutoScalingUpdate - Undocumented member.

$sel:replicaGlobalSecondaryIndexUpdates:ReplicaAutoScalingUpdate', replicaAutoScalingUpdate_replicaGlobalSecondaryIndexUpdates - Represents the auto scaling settings of global secondary indexes that will be modified.

$sel:regionName:ReplicaAutoScalingUpdate', replicaAutoScalingUpdate_regionName - The Region where the replica exists.

replicaAutoScalingUpdate_replicaProvisionedReadCapacityAutoScalingUpdate :: Lens' ReplicaAutoScalingUpdate (Maybe AutoScalingSettingsUpdate)

Undocumented member.

replicaAutoScalingUpdate_replicaGlobalSecondaryIndexUpdates :: Lens' ReplicaAutoScalingUpdate (Maybe [ReplicaGlobalSecondaryIndexAutoScalingUpdate])

Represents the auto scaling settings of global secondary indexes that will be modified.

replicaAutoScalingUpdate_regionName :: Lens' ReplicaAutoScalingUpdate Text

The Region where the replica exists.