Network.AWS.DynamoDB.Types.ReplicaDescription

Description

Documentation

data ReplicaDescription

Contains the details of the replica.

See: newReplicaDescription smart constructor.

Constructors

ReplicaDescription'

Fields

regionName :: Maybe Text

The name of the Region.

globalSecondaryIndexes :: Maybe [ReplicaGlobalSecondaryIndexDescription]

Replica-specific global secondary index settings.

provisionedThroughputOverride :: Maybe ProvisionedThroughputOverride

Replica-specific provisioned throughput. If not described, uses the source table's provisioned throughput settings.

kmsMasterKeyId :: Maybe Text

The AWS KMS customer master key (CMK) of the replica that will be used for AWS KMS encryption.

replicaStatusDescription :: Maybe Text

Detailed information about the replica status.

replicaStatusPercentProgress :: Maybe Text

Specifies the progress of a Create, Update, or Delete action on the replica as a percentage.

replicaStatus :: Maybe ReplicaStatus

The current state of the replica:

CREATING - The replica is being created.

UPDATING - The replica is being updated.

DELETING - The replica is being deleted.

ACTIVE - The replica is ready for use.

REGION_DISABLED - The replica is inaccessible because the AWS Region has been disabled.

If the AWS Region remains inaccessible for more than 20 hours, DynamoDB will remove this replica from the replication group. The replica will not be deleted and replication will stop from and to this region.

INACCESSIBLE_ENCRYPTION_CREDENTIALS - The AWS KMS key used to encrypt the table is inaccessible.

If the AWS KMS key remains inaccessible for more than 20 hours, DynamoDB will remove this replica from the replication group. The replica will not be deleted and replication will stop from and to this region.

replicaInaccessibleDateTime :: Maybe POSIX

The time at which the replica was first detected as inaccessible. To determine cause of inaccessibility check the ReplicaStatus property.

newReplicaDescription :: ReplicaDescription

Create a value of ReplicaDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:regionName:ReplicaDescription', replicaDescription_regionName - The name of the Region.

$sel:globalSecondaryIndexes:ReplicaDescription', replicaDescription_globalSecondaryIndexes - Replica-specific global secondary index settings.

$sel:provisionedThroughputOverride:ReplicaDescription', replicaDescription_provisionedThroughputOverride - Replica-specific provisioned throughput. If not described, uses the source table's provisioned throughput settings.

$sel:kmsMasterKeyId:ReplicaDescription', replicaDescription_kmsMasterKeyId - The AWS KMS customer master key (CMK) of the replica that will be used for AWS KMS encryption.

$sel:replicaStatusDescription:ReplicaDescription', replicaDescription_replicaStatusDescription - Detailed information about the replica status.

$sel:replicaStatusPercentProgress:ReplicaDescription', replicaDescription_replicaStatusPercentProgress - Specifies the progress of a Create, Update, or Delete action on the replica as a percentage.

$sel:replicaStatus:ReplicaDescription', replicaDescription_replicaStatus - The current state of the replica:

CREATING - The replica is being created.

UPDATING - The replica is being updated.

DELETING - The replica is being deleted.

ACTIVE - The replica is ready for use.

REGION_DISABLED - The replica is inaccessible because the AWS Region has been disabled.

If the AWS Region remains inaccessible for more than 20 hours, DynamoDB will remove this replica from the replication group. The replica will not be deleted and replication will stop from and to this region.

INACCESSIBLE_ENCRYPTION_CREDENTIALS - The AWS KMS key used to encrypt the table is inaccessible.

If the AWS KMS key remains inaccessible for more than 20 hours, DynamoDB will remove this replica from the replication group. The replica will not be deleted and replication will stop from and to this region.

$sel:replicaInaccessibleDateTime:ReplicaDescription', replicaDescription_replicaInaccessibleDateTime - The time at which the replica was first detected as inaccessible. To determine cause of inaccessibility check the ReplicaStatus property.

replicaDescription_regionName :: Lens' ReplicaDescription (Maybe Text)

The name of the Region.

replicaDescription_globalSecondaryIndexes :: Lens' ReplicaDescription (Maybe [ReplicaGlobalSecondaryIndexDescription])

Replica-specific global secondary index settings.

replicaDescription_provisionedThroughputOverride :: Lens' ReplicaDescription (Maybe ProvisionedThroughputOverride)

Replica-specific provisioned throughput. If not described, uses the source table's provisioned throughput settings.

replicaDescription_kmsMasterKeyId :: Lens' ReplicaDescription (Maybe Text)

The AWS KMS customer master key (CMK) of the replica that will be used for AWS KMS encryption.

replicaDescription_replicaStatusDescription :: Lens' ReplicaDescription (Maybe Text)

Detailed information about the replica status.

replicaDescription_replicaStatusPercentProgress :: Lens' ReplicaDescription (Maybe Text)

Specifies the progress of a Create, Update, or Delete action on the replica as a percentage.

replicaDescription_replicaStatus :: Lens' ReplicaDescription (Maybe ReplicaStatus)

The current state of the replica:

CREATING - The replica is being created.

UPDATING - The replica is being updated.

DELETING - The replica is being deleted.

ACTIVE - The replica is ready for use.

REGION_DISABLED - The replica is inaccessible because the AWS Region has been disabled.

If the AWS Region remains inaccessible for more than 20 hours, DynamoDB will remove this replica from the replication group. The replica will not be deleted and replication will stop from and to this region.

INACCESSIBLE_ENCRYPTION_CREDENTIALS - The AWS KMS key used to encrypt the table is inaccessible.

If the AWS KMS key remains inaccessible for more than 20 hours, DynamoDB will remove this replica from the replication group. The replica will not be deleted and replication will stop from and to this region.

replicaDescription_replicaInaccessibleDateTime :: Lens' ReplicaDescription (Maybe UTCTime)

The time at which the replica was first detected as inaccessible. To determine cause of inaccessibility check the ReplicaStatus property.