Network.AWS.DynamoDB.Types.ReplicaSettingsDescription

Description

Documentation

data ReplicaSettingsDescription

Represents the properties of a replica.

See: newReplicaSettingsDescription smart constructor.

Constructors

ReplicaSettingsDescription'

Fields

replicaBillingModeSummary :: Maybe BillingModeSummary

The read/write capacity mode of the replica.

replicaProvisionedReadCapacityAutoScalingSettings :: Maybe AutoScalingSettingsDescription

Auto scaling settings for a global table replica's read capacity units.

replicaProvisionedWriteCapacityAutoScalingSettings :: Maybe AutoScalingSettingsDescription

Auto scaling settings for a global table replica's write capacity units.

replicaGlobalSecondaryIndexSettings :: Maybe [ReplicaGlobalSecondaryIndexSettingsDescription]

Replica global secondary index settings for the global table.

replicaStatus :: Maybe ReplicaStatus

The current state of the Region:

CREATING - The Region is being created.

UPDATING - The Region is being updated.

DELETING - The Region is being deleted.

ACTIVE - The Region is ready for use.

replicaProvisionedWriteCapacityUnits :: Maybe Natural

The maximum number of writes consumed per second before DynamoDB returns a ThrottlingException. For more information, see Specifying Read and Write Requirements in the Amazon DynamoDB Developer Guide.

replicaProvisionedReadCapacityUnits :: Maybe Natural

The maximum number of strongly consistent reads consumed per second before DynamoDB returns a ThrottlingException. For more information, see Specifying Read and Write Requirements in the Amazon DynamoDB Developer Guide.

regionName :: Text

The Region name of the replica.

newReplicaSettingsDescription

Arguments

:: Text
$sel:regionName:ReplicaSettingsDescription'

-> ReplicaSettingsDescription

Create a value of ReplicaSettingsDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:replicaBillingModeSummary:ReplicaSettingsDescription', replicaSettingsDescription_replicaBillingModeSummary - The read/write capacity mode of the replica.

$sel:replicaProvisionedReadCapacityAutoScalingSettings:ReplicaSettingsDescription', replicaSettingsDescription_replicaProvisionedReadCapacityAutoScalingSettings - Auto scaling settings for a global table replica's read capacity units.

$sel:replicaProvisionedWriteCapacityAutoScalingSettings:ReplicaSettingsDescription', replicaSettingsDescription_replicaProvisionedWriteCapacityAutoScalingSettings - Auto scaling settings for a global table replica's write capacity units.

$sel:replicaGlobalSecondaryIndexSettings:ReplicaSettingsDescription', replicaSettingsDescription_replicaGlobalSecondaryIndexSettings - Replica global secondary index settings for the global table.

$sel:replicaStatus:ReplicaSettingsDescription', replicaSettingsDescription_replicaStatus - The current state of the Region:

CREATING - The Region is being created.

UPDATING - The Region is being updated.

DELETING - The Region is being deleted.

ACTIVE - The Region is ready for use.

$sel:replicaProvisionedWriteCapacityUnits:ReplicaSettingsDescription', replicaSettingsDescription_replicaProvisionedWriteCapacityUnits - The maximum number of writes consumed per second before DynamoDB returns a ThrottlingException. For more information, see Specifying Read and Write Requirements in the Amazon DynamoDB Developer Guide.

$sel:replicaProvisionedReadCapacityUnits:ReplicaSettingsDescription', replicaSettingsDescription_replicaProvisionedReadCapacityUnits - The maximum number of strongly consistent reads consumed per second before DynamoDB returns a ThrottlingException. For more information, see Specifying Read and Write Requirements in the Amazon DynamoDB Developer Guide.

$sel:regionName:ReplicaSettingsDescription', replicaSettingsDescription_regionName - The Region name of the replica.

replicaSettingsDescription_replicaBillingModeSummary :: Lens' ReplicaSettingsDescription (Maybe BillingModeSummary)

The read/write capacity mode of the replica.

replicaSettingsDescription_replicaProvisionedReadCapacityAutoScalingSettings :: Lens' ReplicaSettingsDescription (Maybe AutoScalingSettingsDescription)

Auto scaling settings for a global table replica's read capacity units.

replicaSettingsDescription_replicaProvisionedWriteCapacityAutoScalingSettings :: Lens' ReplicaSettingsDescription (Maybe AutoScalingSettingsDescription)

Auto scaling settings for a global table replica's write capacity units.

replicaSettingsDescription_replicaGlobalSecondaryIndexSettings :: Lens' ReplicaSettingsDescription (Maybe [ReplicaGlobalSecondaryIndexSettingsDescription])

Replica global secondary index settings for the global table.

replicaSettingsDescription_replicaStatus :: Lens' ReplicaSettingsDescription (Maybe ReplicaStatus)

The current state of the Region:

CREATING - The Region is being created.

UPDATING - The Region is being updated.

DELETING - The Region is being deleted.

ACTIVE - The Region is ready for use.

replicaSettingsDescription_replicaProvisionedWriteCapacityUnits :: Lens' ReplicaSettingsDescription (Maybe Natural)

The maximum number of writes consumed per second before DynamoDB returns a ThrottlingException. For more information, see Specifying Read and Write Requirements in the Amazon DynamoDB Developer Guide.

replicaSettingsDescription_replicaProvisionedReadCapacityUnits :: Lens' ReplicaSettingsDescription (Maybe Natural)

The maximum number of strongly consistent reads consumed per second before DynamoDB returns a ThrottlingException. For more information, see Specifying Read and Write Requirements in the Amazon DynamoDB Developer Guide.

replicaSettingsDescription_regionName :: Lens' ReplicaSettingsDescription Text

The Region name of the replica.