Network.AWS.DynamoDB.Types.SSESpecification

Description

Documentation

data SSESpecification

Represents the settings used to enable server-side encryption.

See: newSSESpecification smart constructor.

Constructors

SSESpecification'

Fields

kmsMasterKeyId :: Maybe Text

The AWS KMS customer master key (CMK) that should be used for the AWS KMS encryption. To specify a CMK, use its key ID, Amazon Resource Name (ARN), alias name, or alias ARN. Note that you should only provide this parameter if the key is different from the default DynamoDB customer master key alias/aws/dynamodb.

enabled :: Maybe Bool

Indicates whether server-side encryption is done using an AWS managed CMK or an AWS owned CMK. If enabled (true), server-side encryption type is set to KMS and an AWS managed CMK is used (AWS KMS charges apply). If disabled (false) or not specified, server-side encryption is set to AWS owned CMK.

sSEType :: Maybe SSEType

Server-side encryption type. The only supported value is:

KMS - Server-side encryption that uses AWS Key Management Service. The key is stored in your account and is managed by AWS KMS (AWS KMS charges apply).

newSSESpecification :: SSESpecification

Create a value of SSESpecification with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:kmsMasterKeyId:SSESpecification', sSESpecification_kmsMasterKeyId - The AWS KMS customer master key (CMK) that should be used for the AWS KMS encryption. To specify a CMK, use its key ID, Amazon Resource Name (ARN), alias name, or alias ARN. Note that you should only provide this parameter if the key is different from the default DynamoDB customer master key alias/aws/dynamodb.

$sel:enabled:SSESpecification', sSESpecification_enabled - Indicates whether server-side encryption is done using an AWS managed CMK or an AWS owned CMK. If enabled (true), server-side encryption type is set to KMS and an AWS managed CMK is used (AWS KMS charges apply). If disabled (false) or not specified, server-side encryption is set to AWS owned CMK.

$sel:sSEType:SSESpecification', sSESpecification_sSEType - Server-side encryption type. The only supported value is:

KMS - Server-side encryption that uses AWS Key Management Service. The key is stored in your account and is managed by AWS KMS (AWS KMS charges apply).

sSESpecification_kmsMasterKeyId :: Lens' SSESpecification (Maybe Text)

The AWS KMS customer master key (CMK) that should be used for the AWS KMS encryption. To specify a CMK, use its key ID, Amazon Resource Name (ARN), alias name, or alias ARN. Note that you should only provide this parameter if the key is different from the default DynamoDB customer master key alias/aws/dynamodb.

sSESpecification_enabled :: Lens' SSESpecification (Maybe Bool)

Indicates whether server-side encryption is done using an AWS managed CMK or an AWS owned CMK. If enabled (true), server-side encryption type is set to KMS and an AWS managed CMK is used (AWS KMS charges apply). If disabled (false) or not specified, server-side encryption is set to AWS owned CMK.

sSESpecification_sSEType :: Lens' SSESpecification (Maybe SSEType)

Server-side encryption type. The only supported value is:

KMS - Server-side encryption that uses AWS Key Management Service. The key is stored in your account and is managed by AWS KMS (AWS KMS charges apply).