Network.AWS.DynamoDB.Types.SourceTableDetails

Description

Documentation

data SourceTableDetails

Contains the details of the table when the backup was created.

See: newSourceTableDetails smart constructor.

Constructors

SourceTableDetails'

Fields

tableArn :: Maybe Text

ARN of the table for which backup was created.

billingMode :: Maybe BillingMode

Controls how you are charged for read and write throughput and how you manage capacity. This setting can be changed later.

PROVISIONED - Sets the read/write capacity mode to PROVISIONED. We recommend using PROVISIONED for predictable workloads.

PAY_PER_REQUEST - Sets the read/write capacity mode to PAY_PER_REQUEST. We recommend using PAY_PER_REQUEST for unpredictable workloads.

tableSizeBytes :: Maybe Integer

Size of the table in bytes. Note that this is an approximate value.

itemCount :: Maybe Natural

Number of items in the table. Note that this is an approximate value.

tableName :: Text

The name of the table for which the backup was created.

tableId :: Text

Unique identifier for the table for which the backup was created.

keySchema :: NonEmpty KeySchemaElement

Schema of the table.

tableCreationDateTime :: POSIX

Time when the source table was created.

provisionedThroughput :: ProvisionedThroughput

Read IOPs and Write IOPS on the table when the backup was created.

newSourceTableDetails

Arguments

:: Text
$sel:tableName:SourceTableDetails'

-> Text
$sel:tableId:SourceTableDetails'

-> NonEmpty KeySchemaElement
$sel:keySchema:SourceTableDetails'

-> UTCTime
$sel:tableCreationDateTime:SourceTableDetails'

-> ProvisionedThroughput
$sel:provisionedThroughput:SourceTableDetails'

-> SourceTableDetails

Create a value of SourceTableDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tableArn:SourceTableDetails', sourceTableDetails_tableArn - ARN of the table for which backup was created.

$sel:billingMode:SourceTableDetails', sourceTableDetails_billingMode - Controls how you are charged for read and write throughput and how you manage capacity. This setting can be changed later.

PROVISIONED - Sets the read/write capacity mode to PROVISIONED. We recommend using PROVISIONED for predictable workloads.

PAY_PER_REQUEST - Sets the read/write capacity mode to PAY_PER_REQUEST. We recommend using PAY_PER_REQUEST for unpredictable workloads.

$sel:tableSizeBytes:SourceTableDetails', sourceTableDetails_tableSizeBytes - Size of the table in bytes. Note that this is an approximate value.

$sel:itemCount:SourceTableDetails', sourceTableDetails_itemCount - Number of items in the table. Note that this is an approximate value.

$sel:tableName:SourceTableDetails', sourceTableDetails_tableName - The name of the table for which the backup was created.

$sel:tableId:SourceTableDetails', sourceTableDetails_tableId - Unique identifier for the table for which the backup was created.

$sel:keySchema:SourceTableDetails', sourceTableDetails_keySchema - Schema of the table.

$sel:tableCreationDateTime:SourceTableDetails', sourceTableDetails_tableCreationDateTime - Time when the source table was created.

$sel:provisionedThroughput:SourceTableDetails', sourceTableDetails_provisionedThroughput - Read IOPs and Write IOPS on the table when the backup was created.

sourceTableDetails_tableArn :: Lens' SourceTableDetails (Maybe Text)

ARN of the table for which backup was created.

sourceTableDetails_billingMode :: Lens' SourceTableDetails (Maybe BillingMode)

Controls how you are charged for read and write throughput and how you manage capacity. This setting can be changed later.

PROVISIONED - Sets the read/write capacity mode to PROVISIONED. We recommend using PROVISIONED for predictable workloads.

PAY_PER_REQUEST - Sets the read/write capacity mode to PAY_PER_REQUEST. We recommend using PAY_PER_REQUEST for unpredictable workloads.

sourceTableDetails_tableSizeBytes :: Lens' SourceTableDetails (Maybe Integer)

Size of the table in bytes. Note that this is an approximate value.

sourceTableDetails_itemCount :: Lens' SourceTableDetails (Maybe Natural)

Number of items in the table. Note that this is an approximate value.

sourceTableDetails_tableName :: Lens' SourceTableDetails Text

The name of the table for which the backup was created.

sourceTableDetails_tableId :: Lens' SourceTableDetails Text

Unique identifier for the table for which the backup was created.

sourceTableDetails_keySchema :: Lens' SourceTableDetails (NonEmpty KeySchemaElement)

Schema of the table.

sourceTableDetails_tableCreationDateTime :: Lens' SourceTableDetails UTCTime

Time when the source table was created.

sourceTableDetails_provisionedThroughput :: Lens' SourceTableDetails ProvisionedThroughput

Read IOPs and Write IOPS on the table when the backup was created.