Network.AWS.DynamoDB.Types.WriteRequest

Description

Documentation

data WriteRequest

Represents an operation to perform - either DeleteItem or PutItem. You can only request one of these operations, not both, in a single WriteRequest. If you do need to perform both of these operations, you need to provide two separate WriteRequest objects.

See: newWriteRequest smart constructor.

Constructors

WriteRequest'

Fields

deleteRequest :: Maybe DeleteRequest

A request to perform a DeleteItem operation.

putRequest :: Maybe PutRequest

A request to perform a PutItem operation.

newWriteRequest :: WriteRequest

Create a value of WriteRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deleteRequest:WriteRequest', writeRequest_deleteRequest - A request to perform a DeleteItem operation.

$sel:putRequest:WriteRequest', writeRequest_putRequest - A request to perform a PutItem operation.

writeRequest_deleteRequest :: Lens' WriteRequest (Maybe DeleteRequest)

A request to perform a DeleteItem operation.

writeRequest_putRequest :: Lens' WriteRequest (Maybe PutRequest)

A request to perform a PutItem operation.