Network.AWS.DynamoDB.UpdateTableReplicaAutoScaling

Description

Updates auto scaling settings on your global tables at once.

This operation only applies to Version 2019.11.21 of global tables.

Creating a Request

data UpdateTableReplicaAutoScaling

See: newUpdateTableReplicaAutoScaling smart constructor.

Constructors

UpdateTableReplicaAutoScaling'

Fields

provisionedWriteCapacityAutoScalingUpdate :: Maybe AutoScalingSettingsUpdate

globalSecondaryIndexUpdates :: Maybe (NonEmpty GlobalSecondaryIndexAutoScalingUpdate)

Represents the auto scaling settings of the global secondary indexes of the replica to be updated.

replicaUpdates :: Maybe (NonEmpty ReplicaAutoScalingUpdate)

Represents the auto scaling settings of replicas of the table that will be modified.

tableName :: Text

The name of the global table to be updated.

newUpdateTableReplicaAutoScaling

Arguments

:: Text
$sel:tableName:UpdateTableReplicaAutoScaling'

-> UpdateTableReplicaAutoScaling

Create a value of UpdateTableReplicaAutoScaling with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:provisionedWriteCapacityAutoScalingUpdate:UpdateTableReplicaAutoScaling', updateTableReplicaAutoScaling_provisionedWriteCapacityAutoScalingUpdate - Undocumented member.

$sel:globalSecondaryIndexUpdates:UpdateTableReplicaAutoScaling', updateTableReplicaAutoScaling_globalSecondaryIndexUpdates - Represents the auto scaling settings of the global secondary indexes of the replica to be updated.

$sel:replicaUpdates:UpdateTableReplicaAutoScaling', updateTableReplicaAutoScaling_replicaUpdates - Represents the auto scaling settings of replicas of the table that will be modified.

$sel:tableName:UpdateTableReplicaAutoScaling', updateTableReplicaAutoScaling_tableName - The name of the global table to be updated.

Request Lenses

updateTableReplicaAutoScaling_provisionedWriteCapacityAutoScalingUpdate :: Lens' UpdateTableReplicaAutoScaling (Maybe AutoScalingSettingsUpdate)

Undocumented member.

updateTableReplicaAutoScaling_globalSecondaryIndexUpdates :: Lens' UpdateTableReplicaAutoScaling (Maybe (NonEmpty GlobalSecondaryIndexAutoScalingUpdate))

Represents the auto scaling settings of the global secondary indexes of the replica to be updated.

updateTableReplicaAutoScaling_replicaUpdates :: Lens' UpdateTableReplicaAutoScaling (Maybe (NonEmpty ReplicaAutoScalingUpdate))

Represents the auto scaling settings of replicas of the table that will be modified.

updateTableReplicaAutoScaling_tableName :: Lens' UpdateTableReplicaAutoScaling Text

The name of the global table to be updated.

Destructuring the Response

data UpdateTableReplicaAutoScalingResponse

See: newUpdateTableReplicaAutoScalingResponse smart constructor.

Constructors

UpdateTableReplicaAutoScalingResponse'

Fields

tableAutoScalingDescription :: Maybe TableAutoScalingDescription

Returns information about the auto scaling settings of a table with replicas.

httpStatus :: Int

The response's http status code.

newUpdateTableReplicaAutoScalingResponse

Arguments

:: Int
$sel:httpStatus:UpdateTableReplicaAutoScalingResponse'

-> UpdateTableReplicaAutoScalingResponse

Create a value of UpdateTableReplicaAutoScalingResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tableAutoScalingDescription:UpdateTableReplicaAutoScalingResponse', updateTableReplicaAutoScalingResponse_tableAutoScalingDescription - Returns information about the auto scaling settings of a table with replicas.

$sel:httpStatus:UpdateTableReplicaAutoScalingResponse', updateTableReplicaAutoScalingResponse_httpStatus - The response's http status code.

Response Lenses

updateTableReplicaAutoScalingResponse_tableAutoScalingDescription :: Lens' UpdateTableReplicaAutoScalingResponse (Maybe TableAutoScalingDescription)

Returns information about the auto scaling settings of a table with replicas.

updateTableReplicaAutoScalingResponse_httpStatus :: Lens' UpdateTableReplicaAutoScalingResponse Int

The response's http status code.