Network.AWS.DynamoDB.V2012_08_10.GetItem

Description

The GetItem operation returns a set of attributes for the item with the given primary key. If there is no matching item, GetItem does not return any data. GetItem provides an eventually consistent read by default. If your application requires a strongly consistent read, set ConsistentRead to true. Although a strongly consistent read might take more time than an eventually consistent read, it always returns the last updated value. Retrieve Item Attributes This example retrieves three attributes from the Thread table. In the response, the ConsumedCapacityUnits value is 1, because ConsistentRead is set to true. If ConsistentRead had been set to false (or not specified) for the same request, an eventually consistent read would have been used and ConsumedCapacityUnits would have been 0.5. { ConsumedCapacity: { CapacityUnits: 1, TableName: Thread }, Item: { Tags: { SS: [Update,"Multiple Items",HelpMe] }, LastPostDateTime: { S: "201303190436" }, Message: { S: "I want to update multiple items in a single API call. What's the best way to do that?" } } }.

Documentation

getItem

Arguments

:: HashMap Text AttributeValue
_giiKey

-> Text
_giiTableName

-> GetItem

Minimum specification for a GetItem request.

data GetItem

Constructors

GetItem

Fields

_giiKey :: HashMap Text AttributeValue

A map of attribute names to AttributeValue objects, representing the primary key of the item to retrieve.

_giiTableName :: Text

The name of the table containing the requested item.

_giiAttributesToGet :: Maybe [Text]

The names of one or more attributes to retrieve. If no attribute names are specified, then all attributes will be returned. If any of the requested attributes are not found, they will not appear in the result.

_giiConsistentRead :: Maybe Bool

If set to true, then the operation uses strongly consistent reads; otherwise, eventually consistent reads are used.

_giiReturnConsumedCapacity :: Maybe ReturnConsumedCapacity

If set to TOTAL, the response includes ConsumedCapacity data for tables and indexes. If set to INDEXES, the repsonse includes ConsumedCapacity for indexes. If set to NONE (the default), ConsumedCapacity is not included in the response.

giiTableName :: Lens' GetItem Text

giiReturnConsumedCapacity :: Lens' GetItem (Maybe ReturnConsumedCapacity)

giiKey :: Lens' GetItem (HashMap Text AttributeValue)

giiConsistentRead :: Lens' GetItem (Maybe Bool)

giiAttributesToGet :: Lens' GetItem (Maybe [Text])

data GetItemResponse

Constructors

GetItemResponse

Fields

_gioItem :: HashMap Text AttributeValue

A map of attribute names to AttributeValue objects, as specified by AttributesToGet.

_gioConsumedCapacity :: Maybe ConsumedCapacity

Represents the capacity units consumed by an operation. The data returned includes the total provisioned throughput consumed, along with statistics for the table and any indexes involved in the operation. ConsumedCapacity is only returned if it was asked for in the request. For more information, see Provisioned Throughput in the Amazon DynamoDB Developer Guide.

gioItem :: Lens' GetItemResponse (HashMap Text AttributeValue)

gioConsumedCapacity :: Lens' GetItemResponse (Maybe ConsumedCapacity)