Network.AWS.DynamoDBStreams.Types

Description

Service Configuration

defaultService :: Service

API version 2012-08-10 of the Amazon DynamoDB Streams SDK configuration.

Errors

_TrimmedDataAccessException :: AsError a => Getting (First ServiceError) a ServiceError

The operation attempted to read past the oldest stream record in a shard.

In DynamoDB Streams, there is a 24 hour limit on data retention. Stream records whose age exceeds this limit are subject to removal (trimming) from the stream. You might receive a TrimmedDataAccessException if:

You request a shard iterator with a sequence number older than the trim point (24 hours).

You obtain a shard iterator, but before you use the iterator in a GetRecords request, a stream record in the shard exceeds the 24 hour period and is trimmed. This causes the iterator to access a record that no longer exists.

_ExpiredIteratorException :: AsError a => Getting (First ServiceError) a ServiceError

The shard iterator has expired and can no longer be used to retrieve stream records. A shard iterator expires 15 minutes after it is retrieved using the GetShardIterator action.

_InternalServerError :: AsError a => Getting (First ServiceError) a ServiceError

An error occurred on the server side.

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

There is no limit to the number of daily on-demand backups that can be taken.

Up to 50 simultaneous table operations are allowed per account. These operations include CreateTable, UpdateTable, DeleteTable,UpdateTimeToLive, RestoreTableFromBackup, and RestoreTableToPointInTime.

The only exception is when you are creating a table with one or more secondary indexes. You can have up to 25 such requests running at a time; however, if the table or index specifications are complex, DynamoDB might temporarily reduce the number of concurrent operations.

There is a soft account quota of 256 tables.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The operation tried to access a nonexistent table or index. The resource might not be specified correctly, or its status might not be ACTIVE.

KeyType

newtype KeyType

Constructors

KeyType'

Fields

fromKeyType :: Text

Bundled Patterns

pattern KeyType_HASH :: KeyType

pattern KeyType_RANGE :: KeyType

OperationType

newtype OperationType

Constructors

OperationType'

Fields

fromOperationType :: Text

Bundled Patterns

pattern OperationType_INSERT :: OperationType

pattern OperationType_MODIFY :: OperationType

pattern OperationType_REMOVE :: OperationType

ShardIteratorType

newtype ShardIteratorType

Constructors

ShardIteratorType'

Fields

fromShardIteratorType :: Text

Bundled Patterns

pattern ShardIteratorType_AFTER_SEQUENCE_NUMBER :: ShardIteratorType

pattern ShardIteratorType_AT_SEQUENCE_NUMBER :: ShardIteratorType

pattern ShardIteratorType_LATEST :: ShardIteratorType

pattern ShardIteratorType_TRIM_HORIZON :: ShardIteratorType

StreamStatus

newtype StreamStatus

Constructors

StreamStatus'

Fields

fromStreamStatus :: Text

Bundled Patterns

pattern StreamStatus_DISABLED :: StreamStatus

pattern StreamStatus_DISABLING :: StreamStatus

pattern StreamStatus_ENABLED :: StreamStatus

pattern StreamStatus_ENABLING :: StreamStatus

StreamViewType

newtype StreamViewType

Constructors

StreamViewType'

Fields

fromStreamViewType :: Text

Bundled Patterns

pattern StreamViewType_KEYS_ONLY :: StreamViewType

pattern StreamViewType_NEW_AND_OLD_IMAGES :: StreamViewType

pattern StreamViewType_NEW_IMAGE :: StreamViewType

pattern StreamViewType_OLD_IMAGE :: StreamViewType

AttributeValue

data AttributeValue

Represents the data for an attribute.

Each attribute value is described as a name-value pair. The name is the data type, and the value is the data itself.

For more information, see Data Types in the Amazon DynamoDB Developer Guide.

See: newAttributeValue smart constructor.

Constructors

AttributeValue'

Fields

bs :: Maybe [Base64]

An attribute of type Binary Set. For example:

"BS": ["U3Vubnk=", "UmFpbnk=", "U25vd3k="]

n :: Maybe Text

An attribute of type Number. For example:

"N": "123.45"

Numbers are sent across the network to DynamoDB as strings, to maximize compatibility across languages and libraries. However, DynamoDB treats them as number type attributes for mathematical operations.

bool :: Maybe Bool

An attribute of type Boolean. For example:

"BOOL": true

null :: Maybe Bool

An attribute of type Null. For example:

"NULL": true

s :: Maybe Text

An attribute of type String. For example:

"S": "Hello"

m :: Maybe (HashMap Text AttributeValue)

An attribute of type Map. For example:

"M": {"Name": {"S": "Joe"}, "Age": {"N": "35"}}

b :: Maybe Base64

An attribute of type Binary. For example:

"B": "dGhpcyB0ZXh0IGlzIGJhc2U2NC1lbmNvZGVk"

l :: Maybe [AttributeValue]

An attribute of type List. For example:

"L": [ {"S": "Cookies"} , {"S": "Coffee"}, {"N", "3.14159"}]

ss :: Maybe [Text]

An attribute of type String Set. For example:

"SS": ["Giraffe", "Hippo" ,"Zebra"]

ns :: Maybe [Text]

An attribute of type Number Set. For example:

"NS": ["42.2", "-19", "7.5", "3.14"]

Numbers are sent across the network to DynamoDB as strings, to maximize compatibility across languages and libraries. However, DynamoDB treats them as number type attributes for mathematical operations.

newAttributeValue :: AttributeValue

Create a value of AttributeValue with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bs:AttributeValue', attributeValue_bs - An attribute of type Binary Set. For example:

"BS": ["U3Vubnk=", "UmFpbnk=", "U25vd3k="]

$sel:n:AttributeValue', attributeValue_n - An attribute of type Number. For example:

"N": "123.45"

Numbers are sent across the network to DynamoDB as strings, to maximize compatibility across languages and libraries. However, DynamoDB treats them as number type attributes for mathematical operations.

$sel:bool:AttributeValue', attributeValue_bool - An attribute of type Boolean. For example:

"BOOL": true

$sel:null:AttributeValue', attributeValue_null - An attribute of type Null. For example:

"NULL": true

$sel:s:AttributeValue', attributeValue_s - An attribute of type String. For example:

"S": "Hello"

$sel:m:AttributeValue', attributeValue_m - An attribute of type Map. For example:

"M": {"Name": {"S": "Joe"}, "Age": {"N": "35"}}

$sel:b:AttributeValue', attributeValue_b - An attribute of type Binary. For example:

"B": "dGhpcyB0ZXh0IGlzIGJhc2U2NC1lbmNvZGVk"-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:l:AttributeValue', attributeValue_l - An attribute of type List. For example:

"L": [ {"S": "Cookies"} , {"S": "Coffee"}, {"N", "3.14159"}]

$sel:ss:AttributeValue', attributeValue_ss - An attribute of type String Set. For example:

"SS": ["Giraffe", "Hippo" ,"Zebra"]

$sel:ns:AttributeValue', attributeValue_ns - An attribute of type Number Set. For example:

"NS": ["42.2", "-19", "7.5", "3.14"]

Numbers are sent across the network to DynamoDB as strings, to maximize compatibility across languages and libraries. However, DynamoDB treats them as number type attributes for mathematical operations.

attributeValue_bs :: Lens' AttributeValue (Maybe [ByteString])

An attribute of type Binary Set. For example:

"BS": ["U3Vubnk=", "UmFpbnk=", "U25vd3k="]

attributeValue_n :: Lens' AttributeValue (Maybe Text)

An attribute of type Number. For example:

"N": "123.45"

Numbers are sent across the network to DynamoDB as strings, to maximize compatibility across languages and libraries. However, DynamoDB treats them as number type attributes for mathematical operations.

attributeValue_bool :: Lens' AttributeValue (Maybe Bool)

An attribute of type Boolean. For example:

"BOOL": true

attributeValue_null :: Lens' AttributeValue (Maybe Bool)

An attribute of type Null. For example:

"NULL": true

attributeValue_s :: Lens' AttributeValue (Maybe Text)

An attribute of type String. For example:

"S": "Hello"

attributeValue_m :: Lens' AttributeValue (Maybe (HashMap Text AttributeValue))

An attribute of type Map. For example:

"M": {"Name": {"S": "Joe"}, "Age": {"N": "35"}}

attributeValue_b :: Lens' AttributeValue (Maybe ByteString)

An attribute of type Binary. For example:

"B": "dGhpcyB0ZXh0IGlzIGJhc2U2NC1lbmNvZGVk"-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

attributeValue_l :: Lens' AttributeValue (Maybe [AttributeValue])

An attribute of type List. For example:

"L": [ {"S": "Cookies"} , {"S": "Coffee"}, {"N", "3.14159"}]

attributeValue_ss :: Lens' AttributeValue (Maybe [Text])

An attribute of type String Set. For example:

"SS": ["Giraffe", "Hippo" ,"Zebra"]

attributeValue_ns :: Lens' AttributeValue (Maybe [Text])

An attribute of type Number Set. For example:

"NS": ["42.2", "-19", "7.5", "3.14"]

Numbers are sent across the network to DynamoDB as strings, to maximize compatibility across languages and libraries. However, DynamoDB treats them as number type attributes for mathematical operations.

Identity

data Identity

Contains details about the type of identity that made the request.

See: newIdentity smart constructor.

Constructors

Identity'

Fields

principalId :: Maybe Text

A unique identifier for the entity that made the call. For Time To Live, the principalId is "dynamodb.amazonaws.com".

type' :: Maybe Text

The type of the identity. For Time To Live, the type is "Service".

newIdentity :: Identity

Create a value of Identity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:principalId:Identity', identity_principalId - A unique identifier for the entity that made the call. For Time To Live, the principalId is "dynamodb.amazonaws.com".

$sel:type':Identity', identity_type - The type of the identity. For Time To Live, the type is "Service".

identity_principalId :: Lens' Identity (Maybe Text)

A unique identifier for the entity that made the call. For Time To Live, the principalId is "dynamodb.amazonaws.com".

identity_type :: Lens' Identity (Maybe Text)

The type of the identity. For Time To Live, the type is "Service".

KeySchemaElement

data KeySchemaElement

Represents a single element of a key schema. A key schema specifies the attributes that make up the primary key of a table, or the key attributes of an index.

A KeySchemaElement represents exactly one attribute of the primary key. For example, a simple primary key would be represented by one KeySchemaElement (for the partition key). A composite primary key would require one KeySchemaElement for the partition key, and another KeySchemaElement for the sort key.

A KeySchemaElement must be a scalar, top-level attribute (not a nested attribute). The data type must be one of String, Number, or Binary. The attribute cannot be nested within a List or a Map.

See: newKeySchemaElement smart constructor.

Constructors

KeySchemaElement'

Fields

attributeName :: Text

The name of a key attribute.

keyType :: KeyType

The role that this key attribute will assume:

HASH - partition key

RANGE - sort key

The partition key of an item is also known as its hash attribute. The term "hash attribute" derives from DynamoDB's usage of an internal hash function to evenly distribute data items across partitions, based on their partition key values.

The sort key of an item is also known as its range attribute. The term "range attribute" derives from the way DynamoDB stores items with the same partition key physically close together, in sorted order by the sort key value.

newKeySchemaElement

Arguments

:: Text
$sel:attributeName:KeySchemaElement'

-> KeyType
$sel:keyType:KeySchemaElement'

-> KeySchemaElement

Create a value of KeySchemaElement with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributeName:KeySchemaElement', keySchemaElement_attributeName - The name of a key attribute.

$sel:keyType:KeySchemaElement', keySchemaElement_keyType - The role that this key attribute will assume:

HASH - partition key

RANGE - sort key

The partition key of an item is also known as its hash attribute. The term "hash attribute" derives from DynamoDB's usage of an internal hash function to evenly distribute data items across partitions, based on their partition key values.

The sort key of an item is also known as its range attribute. The term "range attribute" derives from the way DynamoDB stores items with the same partition key physically close together, in sorted order by the sort key value.

keySchemaElement_attributeName :: Lens' KeySchemaElement Text

The name of a key attribute.

keySchemaElement_keyType :: Lens' KeySchemaElement KeyType

The role that this key attribute will assume:

HASH - partition key

RANGE - sort key

The partition key of an item is also known as its hash attribute. The term "hash attribute" derives from DynamoDB's usage of an internal hash function to evenly distribute data items across partitions, based on their partition key values.

The sort key of an item is also known as its range attribute. The term "range attribute" derives from the way DynamoDB stores items with the same partition key physically close together, in sorted order by the sort key value.

Record

data Record

A description of a unique event within a stream.

See: newRecord smart constructor.

Constructors

Record'

Fields

userIdentity :: Maybe Identity

Items that are deleted by the Time to Live process after expiration have the following fields:

Records[].userIdentity.type

"Service"

Records[].userIdentity.principalId

"dynamodb.amazonaws.com"

eventSource :: Maybe Text

The AWS service from which the stream record originated. For DynamoDB Streams, this is aws:dynamodb.

eventID :: Maybe Text

A globally unique identifier for the event that was recorded in this stream record.

eventName :: Maybe OperationType

The type of data modification that was performed on the DynamoDB table:

INSERT - a new item was added to the table.

MODIFY - one or more of an existing item's attributes were modified.

REMOVE - the item was deleted from the table

eventVersion :: Maybe Text

The version number of the stream record format. This number is updated whenever the structure of Record is modified.

Client applications must not assume that eventVersion will remain at a particular value, as this number is subject to change at any time. In general, eventVersion will only increase as the low-level DynamoDB Streams API evolves.

dynamodb :: Maybe StreamRecord

The main body of the stream record, containing all of the DynamoDB-specific fields.

awsRegion :: Maybe Text

The region in which the GetRecords request was received.

newRecord :: Record

Create a value of Record with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userIdentity:Record', record_userIdentity - Items that are deleted by the Time to Live process after expiration have the following fields:

Records[].userIdentity.type

"Service"

Records[].userIdentity.principalId

"dynamodb.amazonaws.com"

$sel:eventSource:Record', record_eventSource - The AWS service from which the stream record originated. For DynamoDB Streams, this is aws:dynamodb.

$sel:eventID:Record', record_eventID - A globally unique identifier for the event that was recorded in this stream record.

$sel:eventName:Record', record_eventName - The type of data modification that was performed on the DynamoDB table:

INSERT - a new item was added to the table.

MODIFY - one or more of an existing item's attributes were modified.

REMOVE - the item was deleted from the table

$sel:eventVersion:Record', record_eventVersion - The version number of the stream record format. This number is updated whenever the structure of Record is modified.

Client applications must not assume that eventVersion will remain at a particular value, as this number is subject to change at any time. In general, eventVersion will only increase as the low-level DynamoDB Streams API evolves.

$sel:dynamodb:Record', record_dynamodb - The main body of the stream record, containing all of the DynamoDB-specific fields.

$sel:awsRegion:Record', record_awsRegion - The region in which the GetRecords request was received.

record_userIdentity :: Lens' Record (Maybe Identity)

Items that are deleted by the Time to Live process after expiration have the following fields:

Records[].userIdentity.type

"Service"

Records[].userIdentity.principalId

"dynamodb.amazonaws.com"

record_eventSource :: Lens' Record (Maybe Text)

The AWS service from which the stream record originated. For DynamoDB Streams, this is aws:dynamodb.

record_eventID :: Lens' Record (Maybe Text)

A globally unique identifier for the event that was recorded in this stream record.

record_eventName :: Lens' Record (Maybe OperationType)

The type of data modification that was performed on the DynamoDB table:

INSERT - a new item was added to the table.

MODIFY - one or more of an existing item's attributes were modified.

REMOVE - the item was deleted from the table

record_eventVersion :: Lens' Record (Maybe Text)

The version number of the stream record format. This number is updated whenever the structure of Record is modified.

Client applications must not assume that eventVersion will remain at a particular value, as this number is subject to change at any time. In general, eventVersion will only increase as the low-level DynamoDB Streams API evolves.

record_dynamodb :: Lens' Record (Maybe StreamRecord)

The main body of the stream record, containing all of the DynamoDB-specific fields.

record_awsRegion :: Lens' Record (Maybe Text)

The region in which the GetRecords request was received.

SequenceNumberRange

data SequenceNumberRange

The beginning and ending sequence numbers for the stream records contained within a shard.

See: newSequenceNumberRange smart constructor.

Constructors

SequenceNumberRange'

Fields

startingSequenceNumber :: Maybe Text

The first sequence number for the stream records contained within a shard. String contains numeric characters only.

endingSequenceNumber :: Maybe Text

The last sequence number for the stream records contained within a shard. String contains numeric characters only.

newSequenceNumberRange :: SequenceNumberRange

Create a value of SequenceNumberRange with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:startingSequenceNumber:SequenceNumberRange', sequenceNumberRange_startingSequenceNumber - The first sequence number for the stream records contained within a shard. String contains numeric characters only.

$sel:endingSequenceNumber:SequenceNumberRange', sequenceNumberRange_endingSequenceNumber - The last sequence number for the stream records contained within a shard. String contains numeric characters only.

sequenceNumberRange_startingSequenceNumber :: Lens' SequenceNumberRange (Maybe Text)

The first sequence number for the stream records contained within a shard. String contains numeric characters only.

sequenceNumberRange_endingSequenceNumber :: Lens' SequenceNumberRange (Maybe Text)

The last sequence number for the stream records contained within a shard. String contains numeric characters only.

Shard

data Shard

A uniquely identified group of stream records within a stream.

See: newShard smart constructor.

Constructors

Shard'

Fields

shardId :: Maybe Text

The system-generated identifier for this shard.

sequenceNumberRange :: Maybe SequenceNumberRange

The range of possible sequence numbers for the shard.

parentShardId :: Maybe Text

The shard ID of the current shard's parent.

newShard :: Shard

Create a value of Shard with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:shardId:Shard', shard_shardId - The system-generated identifier for this shard.

$sel:sequenceNumberRange:Shard', shard_sequenceNumberRange - The range of possible sequence numbers for the shard.

$sel:parentShardId:Shard', shard_parentShardId - The shard ID of the current shard's parent.

shard_shardId :: Lens' Shard (Maybe Text)

The system-generated identifier for this shard.

shard_sequenceNumberRange :: Lens' Shard (Maybe SequenceNumberRange)

The range of possible sequence numbers for the shard.

shard_parentShardId :: Lens' Shard (Maybe Text)

The shard ID of the current shard's parent.

Stream

data Stream

Represents all of the data describing a particular stream.

See: newStream smart constructor.

Constructors

Stream'

Fields

tableName :: Maybe Text

The DynamoDB table with which the stream is associated.

streamArn :: Maybe Text

The Amazon Resource Name (ARN) for the stream.

streamLabel :: Maybe Text

A timestamp, in ISO 8601 format, for this stream.

Note that LatestStreamLabel is not a unique identifier for the stream, because it is possible that a stream from another table might have the same timestamp. However, the combination of the following three elements is guaranteed to be unique:

the AWS customer ID.

the table name

the StreamLabel

newStream :: Stream

Create a value of Stream with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tableName:Stream', stream_tableName - The DynamoDB table with which the stream is associated.

$sel:streamArn:Stream', stream_streamArn - The Amazon Resource Name (ARN) for the stream.

$sel:streamLabel:Stream', stream_streamLabel - A timestamp, in ISO 8601 format, for this stream.

Note that LatestStreamLabel is not a unique identifier for the stream, because it is possible that a stream from another table might have the same timestamp. However, the combination of the following three elements is guaranteed to be unique:

the AWS customer ID.

the table name

the StreamLabel

stream_tableName :: Lens' Stream (Maybe Text)

The DynamoDB table with which the stream is associated.

stream_streamArn :: Lens' Stream (Maybe Text)

The Amazon Resource Name (ARN) for the stream.

stream_streamLabel :: Lens' Stream (Maybe Text)

A timestamp, in ISO 8601 format, for this stream.

Note that LatestStreamLabel is not a unique identifier for the stream, because it is possible that a stream from another table might have the same timestamp. However, the combination of the following three elements is guaranteed to be unique:

the AWS customer ID.

the table name

the StreamLabel

StreamDescription

data StreamDescription

Represents all of the data describing a particular stream.

See: newStreamDescription smart constructor.

Constructors

StreamDescription'

Fields

lastEvaluatedShardId :: Maybe Text

The shard ID of the item where the operation stopped, inclusive of the previous result set. Use this value to start a new operation, excluding this value in the new request.

If LastEvaluatedShardId is empty, then the "last page" of results has been processed and there is currently no more data to be retrieved.

If LastEvaluatedShardId is not empty, it does not necessarily mean that there is more data in the result set. The only way to know when you have reached the end of the result set is when LastEvaluatedShardId is empty.

streamViewType :: Maybe StreamViewType

Indicates the format of the records within this stream:

KEYS_ONLY - only the key attributes of items that were modified in the DynamoDB table.

NEW_IMAGE - entire items from the table, as they appeared after they were modified.

OLD_IMAGE - entire items from the table, as they appeared before they were modified.

NEW_AND_OLD_IMAGES - both the new and the old images of the items from the table.

tableName :: Maybe Text

The DynamoDB table with which the stream is associated.

creationRequestDateTime :: Maybe POSIX

The date and time when the request to create this stream was issued.

keySchema :: Maybe (NonEmpty KeySchemaElement)

The key attribute(s) of the stream's DynamoDB table.

streamStatus :: Maybe StreamStatus

Indicates the current status of the stream:

ENABLING - Streams is currently being enabled on the DynamoDB table.

ENABLED - the stream is enabled.

DISABLING - Streams is currently being disabled on the DynamoDB table.

DISABLED - the stream is disabled.

shards :: Maybe [Shard]

The shards that comprise the stream.

streamArn :: Maybe Text

The Amazon Resource Name (ARN) for the stream.

streamLabel :: Maybe Text

A timestamp, in ISO 8601 format, for this stream.

Note that LatestStreamLabel is not a unique identifier for the stream, because it is possible that a stream from another table might have the same timestamp. However, the combination of the following three elements is guaranteed to be unique:

the AWS customer ID.

the table name

the StreamLabel

newStreamDescription :: StreamDescription

Create a value of StreamDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastEvaluatedShardId:StreamDescription', streamDescription_lastEvaluatedShardId - The shard ID of the item where the operation stopped, inclusive of the previous result set. Use this value to start a new operation, excluding this value in the new request.

If LastEvaluatedShardId is empty, then the "last page" of results has been processed and there is currently no more data to be retrieved.

If LastEvaluatedShardId is not empty, it does not necessarily mean that there is more data in the result set. The only way to know when you have reached the end of the result set is when LastEvaluatedShardId is empty.

$sel:streamViewType:StreamDescription', streamDescription_streamViewType - Indicates the format of the records within this stream:

KEYS_ONLY - only the key attributes of items that were modified in the DynamoDB table.

NEW_IMAGE - entire items from the table, as they appeared after they were modified.

OLD_IMAGE - entire items from the table, as they appeared before they were modified.

NEW_AND_OLD_IMAGES - both the new and the old images of the items from the table.

$sel:tableName:StreamDescription', streamDescription_tableName - The DynamoDB table with which the stream is associated.

$sel:creationRequestDateTime:StreamDescription', streamDescription_creationRequestDateTime - The date and time when the request to create this stream was issued.

$sel:keySchema:StreamDescription', streamDescription_keySchema - The key attribute(s) of the stream's DynamoDB table.

$sel:streamStatus:StreamDescription', streamDescription_streamStatus - Indicates the current status of the stream:

ENABLING - Streams is currently being enabled on the DynamoDB table.

ENABLED - the stream is enabled.

DISABLING - Streams is currently being disabled on the DynamoDB table.

DISABLED - the stream is disabled.

$sel:shards:StreamDescription', streamDescription_shards - The shards that comprise the stream.

$sel:streamArn:StreamDescription', streamDescription_streamArn - The Amazon Resource Name (ARN) for the stream.

$sel:streamLabel:StreamDescription', streamDescription_streamLabel - A timestamp, in ISO 8601 format, for this stream.

Note that LatestStreamLabel is not a unique identifier for the stream, because it is possible that a stream from another table might have the same timestamp. However, the combination of the following three elements is guaranteed to be unique:

the AWS customer ID.

the table name

the StreamLabel

streamDescription_lastEvaluatedShardId :: Lens' StreamDescription (Maybe Text)

The shard ID of the item where the operation stopped, inclusive of the previous result set. Use this value to start a new operation, excluding this value in the new request.

If LastEvaluatedShardId is empty, then the "last page" of results has been processed and there is currently no more data to be retrieved.

If LastEvaluatedShardId is not empty, it does not necessarily mean that there is more data in the result set. The only way to know when you have reached the end of the result set is when LastEvaluatedShardId is empty.

streamDescription_streamViewType :: Lens' StreamDescription (Maybe StreamViewType)

Indicates the format of the records within this stream:

KEYS_ONLY - only the key attributes of items that were modified in the DynamoDB table.

NEW_IMAGE - entire items from the table, as they appeared after they were modified.

OLD_IMAGE - entire items from the table, as they appeared before they were modified.

NEW_AND_OLD_IMAGES - both the new and the old images of the items from the table.

streamDescription_tableName :: Lens' StreamDescription (Maybe Text)

The DynamoDB table with which the stream is associated.

streamDescription_creationRequestDateTime :: Lens' StreamDescription (Maybe UTCTime)

The date and time when the request to create this stream was issued.

streamDescription_keySchema :: Lens' StreamDescription (Maybe (NonEmpty KeySchemaElement))

The key attribute(s) of the stream's DynamoDB table.

streamDescription_streamStatus :: Lens' StreamDescription (Maybe StreamStatus)

Indicates the current status of the stream:

ENABLING - Streams is currently being enabled on the DynamoDB table.

ENABLED - the stream is enabled.

DISABLING - Streams is currently being disabled on the DynamoDB table.

DISABLED - the stream is disabled.

streamDescription_shards :: Lens' StreamDescription (Maybe [Shard])

The shards that comprise the stream.

streamDescription_streamArn :: Lens' StreamDescription (Maybe Text)

The Amazon Resource Name (ARN) for the stream.

streamDescription_streamLabel :: Lens' StreamDescription (Maybe Text)

A timestamp, in ISO 8601 format, for this stream.

Note that LatestStreamLabel is not a unique identifier for the stream, because it is possible that a stream from another table might have the same timestamp. However, the combination of the following three elements is guaranteed to be unique:

the AWS customer ID.

the table name

the StreamLabel

StreamRecord

data StreamRecord

A description of a single data modification that was performed on an item in a DynamoDB table.

See: newStreamRecord smart constructor.

Constructors

StreamRecord'

Fields

sequenceNumber :: Maybe Text

The sequence number of the stream record.

streamViewType :: Maybe StreamViewType

The type of data from the modified DynamoDB item that was captured in this stream record:

KEYS_ONLY - only the key attributes of the modified item.

NEW_IMAGE - the entire item, as it appeared after it was modified.

OLD_IMAGE - the entire item, as it appeared before it was modified.

NEW_AND_OLD_IMAGES - both the new and the old item images of the item.

keys :: Maybe (HashMap Text AttributeValue)

The primary key attribute(s) for the DynamoDB item that was modified.

sizeBytes :: Maybe Natural

The size of the stream record, in bytes.

newImage' :: Maybe (HashMap Text AttributeValue)

The item in the DynamoDB table as it appeared after it was modified.

approximateCreationDateTime :: Maybe POSIX

The approximate date and time when the stream record was created, in UNIX epoch time format.

oldImage :: Maybe (HashMap Text AttributeValue)

The item in the DynamoDB table as it appeared before it was modified.

newStreamRecord :: StreamRecord

Create a value of StreamRecord with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sequenceNumber:StreamRecord', streamRecord_sequenceNumber - The sequence number of the stream record.

$sel:streamViewType:StreamRecord', streamRecord_streamViewType - The type of data from the modified DynamoDB item that was captured in this stream record:

KEYS_ONLY - only the key attributes of the modified item.

NEW_IMAGE - the entire item, as it appeared after it was modified.

OLD_IMAGE - the entire item, as it appeared before it was modified.

NEW_AND_OLD_IMAGES - both the new and the old item images of the item.

$sel:keys:StreamRecord', streamRecord_keys - The primary key attribute(s) for the DynamoDB item that was modified.

$sel:sizeBytes:StreamRecord', streamRecord_sizeBytes - The size of the stream record, in bytes.

$sel:newImage':StreamRecord', streamRecord_newImage - The item in the DynamoDB table as it appeared after it was modified.

$sel:approximateCreationDateTime:StreamRecord', streamRecord_approximateCreationDateTime - The approximate date and time when the stream record was created, in UNIX epoch time format.

$sel:oldImage:StreamRecord', streamRecord_oldImage - The item in the DynamoDB table as it appeared before it was modified.

streamRecord_sequenceNumber :: Lens' StreamRecord (Maybe Text)

The sequence number of the stream record.

streamRecord_streamViewType :: Lens' StreamRecord (Maybe StreamViewType)

The type of data from the modified DynamoDB item that was captured in this stream record:

KEYS_ONLY - only the key attributes of the modified item.

NEW_IMAGE - the entire item, as it appeared after it was modified.

OLD_IMAGE - the entire item, as it appeared before it was modified.

NEW_AND_OLD_IMAGES - both the new and the old item images of the item.

streamRecord_keys :: Lens' StreamRecord (Maybe (HashMap Text AttributeValue))

The primary key attribute(s) for the DynamoDB item that was modified.

streamRecord_sizeBytes :: Lens' StreamRecord (Maybe Natural)

The size of the stream record, in bytes.

streamRecord_newImage :: Lens' StreamRecord (Maybe (HashMap Text AttributeValue))

The item in the DynamoDB table as it appeared after it was modified.

streamRecord_approximateCreationDateTime :: Lens' StreamRecord (Maybe UTCTime)

The approximate date and time when the stream record was created, in UNIX epoch time format.

streamRecord_oldImage :: Lens' StreamRecord (Maybe (HashMap Text AttributeValue))

The item in the DynamoDB table as it appeared before it was modified.