Network.AWS.DynamoDBStreams.Types.AttributeValue

Description

Documentation

data AttributeValue

Represents the data for an attribute.

Each attribute value is described as a name-value pair. The name is the data type, and the value is the data itself.

For more information, see Data Types in the Amazon DynamoDB Developer Guide.

See: newAttributeValue smart constructor.

Constructors

AttributeValue'

Fields

l :: Maybe [AttributeValue]

An attribute of type List. For example:

"L": [ {"S": "Cookies"} , {"S": "Coffee"}, {"N", "3.14159"}]

ns :: Maybe [Text]

An attribute of type Number Set. For example:

"NS": ["42.2", "-19", "7.5", "3.14"]

Numbers are sent across the network to DynamoDB as strings, to maximize compatibility across languages and libraries. However, DynamoDB treats them as number type attributes for mathematical operations.

m :: Maybe (HashMap Text AttributeValue)

An attribute of type Map. For example:

"M": {"Name": {"S": "Joe"}, "Age": {"N": "35"}}

null :: Maybe Bool

An attribute of type Null. For example:

"NULL": true

n :: Maybe Text

An attribute of type Number. For example:

"N": "123.45"

Numbers are sent across the network to DynamoDB as strings, to maximize compatibility across languages and libraries. However, DynamoDB treats them as number type attributes for mathematical operations.

bs :: Maybe [Base64]

An attribute of type Binary Set. For example:

"BS": ["U3Vubnk=", "UmFpbnk=", "U25vd3k="]

b :: Maybe Base64

An attribute of type Binary. For example:

"B": "dGhpcyB0ZXh0IGlzIGJhc2U2NC1lbmNvZGVk"

ss :: Maybe [Text]

An attribute of type String Set. For example:

"SS": ["Giraffe", "Hippo" ,"Zebra"]

s :: Maybe Text

An attribute of type String. For example:

"S": "Hello"

bool :: Maybe Bool

An attribute of type Boolean. For example:

"BOOL": true

newAttributeValue :: AttributeValue

Create a value of AttributeValue with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:l:AttributeValue', attributeValue_l - An attribute of type List. For example:

"L": [ {"S": "Cookies"} , {"S": "Coffee"}, {"N", "3.14159"}]

$sel:ns:AttributeValue', attributeValue_ns - An attribute of type Number Set. For example:

"NS": ["42.2", "-19", "7.5", "3.14"]

Numbers are sent across the network to DynamoDB as strings, to maximize compatibility across languages and libraries. However, DynamoDB treats them as number type attributes for mathematical operations.

$sel:m:AttributeValue', attributeValue_m - An attribute of type Map. For example:

"M": {"Name": {"S": "Joe"}, "Age": {"N": "35"}}

$sel:null:AttributeValue', attributeValue_null - An attribute of type Null. For example:

"NULL": true

$sel:n:AttributeValue', attributeValue_n - An attribute of type Number. For example:

"N": "123.45"

Numbers are sent across the network to DynamoDB as strings, to maximize compatibility across languages and libraries. However, DynamoDB treats them as number type attributes for mathematical operations.

$sel:bs:AttributeValue', attributeValue_bs - An attribute of type Binary Set. For example:

"BS": ["U3Vubnk=", "UmFpbnk=", "U25vd3k="]

$sel:b:AttributeValue', attributeValue_b - An attribute of type Binary. For example:

"B": "dGhpcyB0ZXh0IGlzIGJhc2U2NC1lbmNvZGVk"-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:ss:AttributeValue', attributeValue_ss - An attribute of type String Set. For example:

"SS": ["Giraffe", "Hippo" ,"Zebra"]

$sel:s:AttributeValue', attributeValue_s - An attribute of type String. For example:

"S": "Hello"

$sel:bool:AttributeValue', attributeValue_bool - An attribute of type Boolean. For example:

"BOOL": true

attributeValue_l :: Lens' AttributeValue (Maybe [AttributeValue])

An attribute of type List. For example:

"L": [ {"S": "Cookies"} , {"S": "Coffee"}, {"N", "3.14159"}]

attributeValue_ns :: Lens' AttributeValue (Maybe [Text])

An attribute of type Number Set. For example:

"NS": ["42.2", "-19", "7.5", "3.14"]

Numbers are sent across the network to DynamoDB as strings, to maximize compatibility across languages and libraries. However, DynamoDB treats them as number type attributes for mathematical operations.

attributeValue_m :: Lens' AttributeValue (Maybe (HashMap Text AttributeValue))

An attribute of type Map. For example:

"M": {"Name": {"S": "Joe"}, "Age": {"N": "35"}}

attributeValue_null :: Lens' AttributeValue (Maybe Bool)

An attribute of type Null. For example:

"NULL": true

attributeValue_n :: Lens' AttributeValue (Maybe Text)

An attribute of type Number. For example:

"N": "123.45"

Numbers are sent across the network to DynamoDB as strings, to maximize compatibility across languages and libraries. However, DynamoDB treats them as number type attributes for mathematical operations.

attributeValue_bs :: Lens' AttributeValue (Maybe [ByteString])

An attribute of type Binary Set. For example:

"BS": ["U3Vubnk=", "UmFpbnk=", "U25vd3k="]

attributeValue_b :: Lens' AttributeValue (Maybe ByteString)

An attribute of type Binary. For example:

"B": "dGhpcyB0ZXh0IGlzIGJhc2U2NC1lbmNvZGVk"-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

attributeValue_ss :: Lens' AttributeValue (Maybe [Text])

An attribute of type String Set. For example:

"SS": ["Giraffe", "Hippo" ,"Zebra"]

attributeValue_s :: Lens' AttributeValue (Maybe Text)

An attribute of type String. For example:

"S": "Hello"

attributeValue_bool :: Lens' AttributeValue (Maybe Bool)

An attribute of type Boolean. For example:

"BOOL": true