Network.AWS.DynamoDBStreams.Types.Record

Description

Documentation

data Record

A description of a unique event within a stream.

See: newRecord smart constructor.

Constructors

Record'

Fields

userIdentity :: Maybe Identity

Items that are deleted by the Time to Live process after expiration have the following fields:

Records[].userIdentity.type

"Service"

Records[].userIdentity.principalId

"dynamodb.amazonaws.com"

eventVersion :: Maybe Text

The version number of the stream record format. This number is updated whenever the structure of Record is modified.

Client applications must not assume that eventVersion will remain at a particular value, as this number is subject to change at any time. In general, eventVersion will only increase as the low-level DynamoDB Streams API evolves.

dynamodb :: Maybe StreamRecord

The main body of the stream record, containing all of the DynamoDB-specific fields.

awsRegion :: Maybe Text

The region in which the GetRecords request was received.

eventName :: Maybe OperationType

The type of data modification that was performed on the DynamoDB table:

INSERT - a new item was added to the table.

MODIFY - one or more of an existing item's attributes were modified.

REMOVE - the item was deleted from the table

eventSource :: Maybe Text

The AWS service from which the stream record originated. For DynamoDB Streams, this is aws:dynamodb.

eventID :: Maybe Text

A globally unique identifier for the event that was recorded in this stream record.

newRecord :: Record

Create a value of Record with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userIdentity:Record', record_userIdentity - Items that are deleted by the Time to Live process after expiration have the following fields:

Records[].userIdentity.type

"Service"

Records[].userIdentity.principalId

"dynamodb.amazonaws.com"

$sel:eventVersion:Record', record_eventVersion - The version number of the stream record format. This number is updated whenever the structure of Record is modified.

Client applications must not assume that eventVersion will remain at a particular value, as this number is subject to change at any time. In general, eventVersion will only increase as the low-level DynamoDB Streams API evolves.

$sel:dynamodb:Record', record_dynamodb - The main body of the stream record, containing all of the DynamoDB-specific fields.

$sel:awsRegion:Record', record_awsRegion - The region in which the GetRecords request was received.

$sel:eventName:Record', record_eventName - The type of data modification that was performed on the DynamoDB table:

INSERT - a new item was added to the table.

MODIFY - one or more of an existing item's attributes were modified.

REMOVE - the item was deleted from the table

$sel:eventSource:Record', record_eventSource - The AWS service from which the stream record originated. For DynamoDB Streams, this is aws:dynamodb.

$sel:eventID:Record', record_eventID - A globally unique identifier for the event that was recorded in this stream record.

record_userIdentity :: Lens' Record (Maybe Identity)

Items that are deleted by the Time to Live process after expiration have the following fields:

Records[].userIdentity.type

"Service"

Records[].userIdentity.principalId

"dynamodb.amazonaws.com"

record_eventVersion :: Lens' Record (Maybe Text)

The version number of the stream record format. This number is updated whenever the structure of Record is modified.

Client applications must not assume that eventVersion will remain at a particular value, as this number is subject to change at any time. In general, eventVersion will only increase as the low-level DynamoDB Streams API evolves.

record_dynamodb :: Lens' Record (Maybe StreamRecord)

The main body of the stream record, containing all of the DynamoDB-specific fields.

record_awsRegion :: Lens' Record (Maybe Text)

The region in which the GetRecords request was received.

record_eventName :: Lens' Record (Maybe OperationType)

The type of data modification that was performed on the DynamoDB table:

INSERT - a new item was added to the table.

MODIFY - one or more of an existing item's attributes were modified.

REMOVE - the item was deleted from the table

record_eventSource :: Lens' Record (Maybe Text)

The AWS service from which the stream record originated. For DynamoDB Streams, this is aws:dynamodb.

record_eventID :: Lens' Record (Maybe Text)

A globally unique identifier for the event that was recorded in this stream record.