Network.AWS.DynamoDBStreams.Types.StreamDescription

Description

Documentation

data StreamDescription

Represents all of the data describing a particular stream.

See: newStreamDescription smart constructor.

Constructors

StreamDescription'

Fields

lastEvaluatedShardId :: Maybe Text

The shard ID of the item where the operation stopped, inclusive of the previous result set. Use this value to start a new operation, excluding this value in the new request.

If LastEvaluatedShardId is empty, then the "last page" of results has been processed and there is currently no more data to be retrieved.

If LastEvaluatedShardId is not empty, it does not necessarily mean that there is more data in the result set. The only way to know when you have reached the end of the result set is when LastEvaluatedShardId is empty.

streamViewType :: Maybe StreamViewType

Indicates the format of the records within this stream:

KEYS_ONLY - only the key attributes of items that were modified in the DynamoDB table.

NEW_IMAGE - entire items from the table, as they appeared after they were modified.

OLD_IMAGE - entire items from the table, as they appeared before they were modified.

NEW_AND_OLD_IMAGES - both the new and the old images of the items from the table.

tableName :: Maybe Text

The DynamoDB table with which the stream is associated.

creationRequestDateTime :: Maybe POSIX

The date and time when the request to create this stream was issued.

keySchema :: Maybe (NonEmpty KeySchemaElement)

The key attribute(s) of the stream's DynamoDB table.

streamStatus :: Maybe StreamStatus

Indicates the current status of the stream:

ENABLING - Streams is currently being enabled on the DynamoDB table.

ENABLED - the stream is enabled.

DISABLING - Streams is currently being disabled on the DynamoDB table.

DISABLED - the stream is disabled.

shards :: Maybe [Shard]

The shards that comprise the stream.

streamArn :: Maybe Text

The Amazon Resource Name (ARN) for the stream.

streamLabel :: Maybe Text

A timestamp, in ISO 8601 format, for this stream.

Note that LatestStreamLabel is not a unique identifier for the stream, because it is possible that a stream from another table might have the same timestamp. However, the combination of the following three elements is guaranteed to be unique:

the AWS customer ID.

the table name

the StreamLabel

newStreamDescription :: StreamDescription

Create a value of StreamDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastEvaluatedShardId:StreamDescription', streamDescription_lastEvaluatedShardId - The shard ID of the item where the operation stopped, inclusive of the previous result set. Use this value to start a new operation, excluding this value in the new request.

If LastEvaluatedShardId is empty, then the "last page" of results has been processed and there is currently no more data to be retrieved.

If LastEvaluatedShardId is not empty, it does not necessarily mean that there is more data in the result set. The only way to know when you have reached the end of the result set is when LastEvaluatedShardId is empty.

$sel:streamViewType:StreamDescription', streamDescription_streamViewType - Indicates the format of the records within this stream:

KEYS_ONLY - only the key attributes of items that were modified in the DynamoDB table.

NEW_IMAGE - entire items from the table, as they appeared after they were modified.

OLD_IMAGE - entire items from the table, as they appeared before they were modified.

NEW_AND_OLD_IMAGES - both the new and the old images of the items from the table.

$sel:tableName:StreamDescription', streamDescription_tableName - The DynamoDB table with which the stream is associated.

$sel:creationRequestDateTime:StreamDescription', streamDescription_creationRequestDateTime - The date and time when the request to create this stream was issued.

$sel:keySchema:StreamDescription', streamDescription_keySchema - The key attribute(s) of the stream's DynamoDB table.

$sel:streamStatus:StreamDescription', streamDescription_streamStatus - Indicates the current status of the stream:

ENABLING - Streams is currently being enabled on the DynamoDB table.

ENABLED - the stream is enabled.

DISABLING - Streams is currently being disabled on the DynamoDB table.

DISABLED - the stream is disabled.

$sel:shards:StreamDescription', streamDescription_shards - The shards that comprise the stream.

$sel:streamArn:StreamDescription', streamDescription_streamArn - The Amazon Resource Name (ARN) for the stream.

$sel:streamLabel:StreamDescription', streamDescription_streamLabel - A timestamp, in ISO 8601 format, for this stream.

Note that LatestStreamLabel is not a unique identifier for the stream, because it is possible that a stream from another table might have the same timestamp. However, the combination of the following three elements is guaranteed to be unique:

the AWS customer ID.

the table name

the StreamLabel

streamDescription_lastEvaluatedShardId :: Lens' StreamDescription (Maybe Text)

The shard ID of the item where the operation stopped, inclusive of the previous result set. Use this value to start a new operation, excluding this value in the new request.

If LastEvaluatedShardId is empty, then the "last page" of results has been processed and there is currently no more data to be retrieved.

If LastEvaluatedShardId is not empty, it does not necessarily mean that there is more data in the result set. The only way to know when you have reached the end of the result set is when LastEvaluatedShardId is empty.

streamDescription_streamViewType :: Lens' StreamDescription (Maybe StreamViewType)

Indicates the format of the records within this stream:

KEYS_ONLY - only the key attributes of items that were modified in the DynamoDB table.

NEW_IMAGE - entire items from the table, as they appeared after they were modified.

OLD_IMAGE - entire items from the table, as they appeared before they were modified.

NEW_AND_OLD_IMAGES - both the new and the old images of the items from the table.

streamDescription_tableName :: Lens' StreamDescription (Maybe Text)

The DynamoDB table with which the stream is associated.

streamDescription_creationRequestDateTime :: Lens' StreamDescription (Maybe UTCTime)

The date and time when the request to create this stream was issued.

streamDescription_keySchema :: Lens' StreamDescription (Maybe (NonEmpty KeySchemaElement))

The key attribute(s) of the stream's DynamoDB table.

streamDescription_streamStatus :: Lens' StreamDescription (Maybe StreamStatus)

Indicates the current status of the stream:

ENABLING - Streams is currently being enabled on the DynamoDB table.

ENABLED - the stream is enabled.

DISABLING - Streams is currently being disabled on the DynamoDB table.

DISABLED - the stream is disabled.

streamDescription_shards :: Lens' StreamDescription (Maybe [Shard])

The shards that comprise the stream.

streamDescription_streamArn :: Lens' StreamDescription (Maybe Text)

The Amazon Resource Name (ARN) for the stream.

streamDescription_streamLabel :: Lens' StreamDescription (Maybe Text)

A timestamp, in ISO 8601 format, for this stream.

Note that LatestStreamLabel is not a unique identifier for the stream, because it is possible that a stream from another table might have the same timestamp. However, the combination of the following three elements is guaranteed to be unique:

the AWS customer ID.

the table name

the StreamLabel