Network.AWS.EBS.Lens

Description

Operations

StartSnapshot

startSnapshot_kmsKeyArn :: Lens' StartSnapshot (Maybe Text)

The Amazon Resource Name (ARN) of the Key Management Service (KMS) key to be used to encrypt the snapshot. If you do not specify a KMS key, the default Amazon Web Services managed KMS key is used.

If you specify a ParentSnapshotId, omit this parameter; the snapshot will be encrypted using the same KMS key that was used to encrypt the parent snapshot.

If Encrypted is set to true, you must specify a KMS key ARN.

startSnapshot_clientToken :: Lens' StartSnapshot (Maybe Text)

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Idempotency ensures that an API request completes only once. With an idempotent request, if the original request completes successfully. The subsequent retries with the same client token return the result from the original successful request and they have no additional effect.

If you do not specify a client token, one is automatically generated by the Amazon Web Services SDK.

For more information, see Idempotency for StartSnapshot API in the Amazon Elastic Compute Cloud User Guide.

startSnapshot_encrypted :: Lens' StartSnapshot (Maybe Bool)

Indicates whether to encrypt the snapshot. To create an encrypted snapshot, specify true. To create an unencrypted snapshot, omit this parameter.

If you specify a value for ParentSnapshotId, omit this parameter.

If you specify true, the snapshot is encrypted using the KMS key specified using the KmsKeyArn parameter. If no value is specified for KmsKeyArn, the default KMS key for your account is used. If no default KMS key has been specified for your account, the Amazon Web Services managed KMS key is used. To set a default KMS key for your account, use ModifyEbsDefaultKmsKeyId.

If your account is enabled for encryption by default, you cannot set this parameter to false. In this case, you can omit this parameter.

For more information, see Using encryption in the Amazon Elastic Compute Cloud User Guide.

startSnapshot_parentSnapshotId :: Lens' StartSnapshot (Maybe Text)

The ID of the parent snapshot. If there is no parent snapshot, or if you are creating the first snapshot for an on-premises volume, omit this parameter.

If your account is enabled for encryption by default, you cannot use an unencrypted snapshot as a parent snapshot. You must first create an encrypted copy of the parent snapshot using CopySnapshot.

startSnapshot_timeout :: Lens' StartSnapshot (Maybe Natural)

The amount of time (in minutes) after which the snapshot is automatically cancelled if:

No blocks are written to the snapshot.

The snapshot is not completed after writing the last block of data.

If no value is specified, the timeout defaults to 60 minutes.

startSnapshot_description :: Lens' StartSnapshot (Maybe Text)

A description for the snapshot.

startSnapshot_tags :: Lens' StartSnapshot (Maybe [Tag])

The tags to apply to the snapshot.

startSnapshot_volumeSize :: Lens' StartSnapshot Natural

The size of the volume, in GiB. The maximum size is 65536 GiB (64 TiB).

startSnapshotResponse_blockSize :: Lens' StartSnapshotResponse (Maybe Int)

The size of the blocks in the snapshot, in bytes.

startSnapshotResponse_status :: Lens' StartSnapshotResponse (Maybe Status)

The status of the snapshot.

startSnapshotResponse_kmsKeyArn :: Lens' StartSnapshotResponse (Maybe Text)

The Amazon Resource Name (ARN) of the Key Management Service (KMS) key used to encrypt the snapshot.

startSnapshotResponse_startTime :: Lens' StartSnapshotResponse (Maybe UTCTime)

The timestamp when the snapshot was created.

startSnapshotResponse_volumeSize :: Lens' StartSnapshotResponse (Maybe Natural)

The size of the volume, in GiB.

startSnapshotResponse_ownerId :: Lens' StartSnapshotResponse (Maybe Text)

The Amazon Web Services account ID of the snapshot owner.

startSnapshotResponse_parentSnapshotId :: Lens' StartSnapshotResponse (Maybe Text)

The ID of the parent snapshot.

startSnapshotResponse_description :: Lens' StartSnapshotResponse (Maybe Text)

The description of the snapshot.

startSnapshotResponse_tags :: Lens' StartSnapshotResponse (Maybe [Tag])

The tags applied to the snapshot. You can specify up to 50 tags per snapshot. For more information, see Tagging your Amazon EC2 resources in the Amazon Elastic Compute Cloud User Guide.

startSnapshotResponse_snapshotId :: Lens' StartSnapshotResponse (Maybe Text)

The ID of the snapshot.

startSnapshotResponse_httpStatus :: Lens' StartSnapshotResponse Int

The response's http status code.

ListSnapshotBlocks

listSnapshotBlocks_nextToken :: Lens' ListSnapshotBlocks (Maybe Text)

The token to request the next page of results.

listSnapshotBlocks_startingBlockIndex :: Lens' ListSnapshotBlocks (Maybe Natural)

The block index from which the list should start. The list in the response will start from this block index or the next valid block index in the snapshot.

listSnapshotBlocks_maxResults :: Lens' ListSnapshotBlocks (Maybe Natural)

The number of results to return.

listSnapshotBlocks_snapshotId :: Lens' ListSnapshotBlocks Text

The ID of the snapshot from which to get block indexes and block tokens.

listSnapshotBlocksResponse_blockSize :: Lens' ListSnapshotBlocksResponse (Maybe Int)

The size of the blocks in the snapshot, in bytes.

listSnapshotBlocksResponse_blocks :: Lens' ListSnapshotBlocksResponse (Maybe [Block])

An array of objects containing information about the blocks.

listSnapshotBlocksResponse_expiryTime :: Lens' ListSnapshotBlocksResponse (Maybe UTCTime)

The time when the BlockToken expires.

listSnapshotBlocksResponse_volumeSize :: Lens' ListSnapshotBlocksResponse (Maybe Natural)

The size of the volume in GB.

listSnapshotBlocksResponse_nextToken :: Lens' ListSnapshotBlocksResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

listSnapshotBlocksResponse_httpStatus :: Lens' ListSnapshotBlocksResponse Int

The response's http status code.

PutSnapshotBlock

putSnapshotBlock_progress :: Lens' PutSnapshotBlock (Maybe Natural)

The progress of the write process, as a percentage.

putSnapshotBlock_snapshotId :: Lens' PutSnapshotBlock Text

The ID of the snapshot.

putSnapshotBlock_blockIndex :: Lens' PutSnapshotBlock Natural

The block index of the block in which to write the data. A block index is a logical index in units of 512 KiB blocks. To identify the block index, divide the logical offset of the data in the logical volume by the block size (logical offset of data/524288). The logical offset of the data must be 512 KiB aligned.

putSnapshotBlock_dataLength :: Lens' PutSnapshotBlock Int

The size of the data to write to the block, in bytes. Currently, the only supported size is 524288 bytes.

Valid values: 524288

putSnapshotBlock_checksum :: Lens' PutSnapshotBlock Text

A Base64-encoded SHA256 checksum of the data. Only SHA256 checksums are supported.

putSnapshotBlock_checksumAlgorithm :: Lens' PutSnapshotBlock ChecksumAlgorithm

The algorithm used to generate the checksum. Currently, the only supported algorithm is SHA256.

putSnapshotBlock_blockData :: Lens' PutSnapshotBlock HashedBody

The data to write to the block.

The block data is not signed as part of the Signature Version 4 signing process. As a result, you must generate and provide a Base64-encoded SHA256 checksum for the block data using the x-amz-Checksum header. Also, you must specify the checksum algorithm using the x-amz-Checksum-Algorithm header. The checksum that you provide is part of the Signature Version 4 signing process. It is validated against a checksum generated by Amazon EBS to ensure the validity and authenticity of the data. If the checksums do not correspond, the request fails. For more information, see Using checksums with the EBS direct APIs in the Amazon Elastic Compute Cloud User Guide.

putSnapshotBlockResponse_checksumAlgorithm :: Lens' PutSnapshotBlockResponse (Maybe ChecksumAlgorithm)

The algorithm used by Amazon EBS to generate the checksum.

putSnapshotBlockResponse_checksum :: Lens' PutSnapshotBlockResponse (Maybe Text)

The SHA256 checksum generated for the block data by Amazon EBS.

putSnapshotBlockResponse_httpStatus :: Lens' PutSnapshotBlockResponse Int

The response's http status code.

ListChangedBlocks

listChangedBlocks_nextToken :: Lens' ListChangedBlocks (Maybe Text)

The token to request the next page of results.

listChangedBlocks_startingBlockIndex :: Lens' ListChangedBlocks (Maybe Natural)

The block index from which the comparison should start.

The list in the response will start from this block index or the next valid block index in the snapshots.

listChangedBlocks_firstSnapshotId :: Lens' ListChangedBlocks (Maybe Text)

The ID of the first snapshot to use for the comparison.

The FirstSnapshotID parameter must be specified with a SecondSnapshotId parameter; otherwise, an error occurs.

listChangedBlocks_maxResults :: Lens' ListChangedBlocks (Maybe Natural)

The number of results to return.

listChangedBlocks_secondSnapshotId :: Lens' ListChangedBlocks Text

The ID of the second snapshot to use for the comparison.

The SecondSnapshotId parameter must be specified with a FirstSnapshotID parameter; otherwise, an error occurs.

listChangedBlocksResponse_blockSize :: Lens' ListChangedBlocksResponse (Maybe Int)

The size of the blocks in the snapshot, in bytes.

listChangedBlocksResponse_expiryTime :: Lens' ListChangedBlocksResponse (Maybe UTCTime)

The time when the BlockToken expires.

listChangedBlocksResponse_volumeSize :: Lens' ListChangedBlocksResponse (Maybe Natural)

The size of the volume in GB.

listChangedBlocksResponse_changedBlocks :: Lens' ListChangedBlocksResponse (Maybe [ChangedBlock])

An array of objects containing information about the changed blocks.

listChangedBlocksResponse_nextToken :: Lens' ListChangedBlocksResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

listChangedBlocksResponse_httpStatus :: Lens' ListChangedBlocksResponse Int

The response's http status code.

CompleteSnapshot

completeSnapshot_checksumAlgorithm :: Lens' CompleteSnapshot (Maybe ChecksumAlgorithm)

The algorithm used to generate the checksum. Currently, the only supported algorithm is SHA256.

completeSnapshot_checksum :: Lens' CompleteSnapshot (Maybe Text)

An aggregated Base-64 SHA256 checksum based on the checksums of each written block.

To generate the aggregated checksum using the linear aggregation method, arrange the checksums for each written block in ascending order of their block index, concatenate them to form a single string, and then generate the checksum on the entire string using the SHA256 algorithm.

completeSnapshot_checksumAggregationMethod :: Lens' CompleteSnapshot (Maybe ChecksumAggregationMethod)

The aggregation method used to generate the checksum. Currently, the only supported aggregation method is LINEAR.

completeSnapshot_snapshotId :: Lens' CompleteSnapshot Text

The ID of the snapshot.

completeSnapshot_changedBlocksCount :: Lens' CompleteSnapshot Natural

The number of blocks that were written to the snapshot.

completeSnapshotResponse_status :: Lens' CompleteSnapshotResponse (Maybe Status)

The status of the snapshot.

completeSnapshotResponse_httpStatus :: Lens' CompleteSnapshotResponse Int

The response's http status code.

GetSnapshotBlock

getSnapshotBlock_snapshotId :: Lens' GetSnapshotBlock Text

The ID of the snapshot containing the block from which to get data.

getSnapshotBlock_blockIndex :: Lens' GetSnapshotBlock Natural

The block index of the block from which to get data.

Obtain the BlockIndex by running the ListChangedBlocks or ListSnapshotBlocks operations.

getSnapshotBlock_blockToken :: Lens' GetSnapshotBlock Text

The block token of the block from which to get data.

Obtain the BlockToken by running the ListChangedBlocks or ListSnapshotBlocks operations.

getSnapshotBlockResponse_dataLength :: Lens' GetSnapshotBlockResponse (Maybe Int)

The size of the data in the block.

getSnapshotBlockResponse_checksumAlgorithm :: Lens' GetSnapshotBlockResponse (Maybe ChecksumAlgorithm)

The algorithm used to generate the checksum for the block, such as SHA256.

getSnapshotBlockResponse_checksum :: Lens' GetSnapshotBlockResponse (Maybe Text)

The checksum generated for the block, which is Base64 encoded.

getSnapshotBlockResponse_httpStatus :: Lens' GetSnapshotBlockResponse Int

The response's http status code.

getSnapshotBlockResponse_blockData :: Lens' GetSnapshotBlockResponse ResponseBody

The data content of the block.

Types

Block

block_blockIndex :: Lens' Block (Maybe Natural)

The block index.

block_blockToken :: Lens' Block (Maybe Text)

The block token for the block index.

ChangedBlock

changedBlock_blockIndex :: Lens' ChangedBlock (Maybe Natural)

The block index.

changedBlock_secondBlockToken :: Lens' ChangedBlock (Maybe Text)

The block token for the block index of the SecondSnapshotId specified in the ListChangedBlocks operation.

changedBlock_firstBlockToken :: Lens' ChangedBlock (Maybe Text)

The block token for the block index of the FirstSnapshotId specified in the ListChangedBlocks operation. This value is absent if the first snapshot does not have the changed block that is on the second snapshot.

Tag

tag_value :: Lens' Tag (Maybe Text)

The value of the tag.

tag_key :: Lens' Tag (Maybe Text)

The key of the tag.