Network.AWS.EBS.Types.ChangedBlock

Description

Documentation

data ChangedBlock

A block of data in an Amazon Elastic Block Store snapshot that is different from another snapshot of the same volume/snapshot lineage.

See: newChangedBlock smart constructor.

Constructors

ChangedBlock'

Fields

blockIndex :: Maybe Natural

The block index.

secondBlockToken :: Maybe Text

The block token for the block index of the SecondSnapshotId specified in the ListChangedBlocks operation.

firstBlockToken :: Maybe Text

The block token for the block index of the FirstSnapshotId specified in the ListChangedBlocks operation. This value is absent if the first snapshot does not have the changed block that is on the second snapshot.

newChangedBlock :: ChangedBlock

Create a value of ChangedBlock with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:blockIndex:ChangedBlock', changedBlock_blockIndex - The block index.

$sel:secondBlockToken:ChangedBlock', changedBlock_secondBlockToken - The block token for the block index of the SecondSnapshotId specified in the ListChangedBlocks operation.

$sel:firstBlockToken:ChangedBlock', changedBlock_firstBlockToken - The block token for the block index of the FirstSnapshotId specified in the ListChangedBlocks operation. This value is absent if the first snapshot does not have the changed block that is on the second snapshot.

changedBlock_blockIndex :: Lens' ChangedBlock (Maybe Natural)

The block index.

changedBlock_secondBlockToken :: Lens' ChangedBlock (Maybe Text)

The block token for the block index of the SecondSnapshotId specified in the ListChangedBlocks operation.

changedBlock_firstBlockToken :: Lens' ChangedBlock (Maybe Text)

The block token for the block index of the FirstSnapshotId specified in the ListChangedBlocks operation. This value is absent if the first snapshot does not have the changed block that is on the second snapshot.