Network.AWS.EC2

Description

Amazon Elastic Compute Cloud provides resizable computing capacity in the Amazon Web Services cloud.

Actions

AllocateAddress

data AllocateAddress

Acquires an Elastic IP address.An Elastic IP address is for use either in the EC2-Classic platform or in a VPC.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-AllocateAddress.html

Constructors

AllocateAddress

Fields

aaDomain :: Maybe AddressDomain

Set to vpc to allocate the address for use with instances in a VPC.

data AllocateAddressResponse

Constructors

AllocateAddressResponse

Fields

aarRequestId :: !Text

The ID of the request.

aarPublicIp :: !Text

The Elastic IP address.

aarDomain :: !AddressDomain

Indicates whether this Elastic IP address is for use with instances in EC2-Classic (standaard) or instances in a VPC (vpc).

aarAllocationId :: Maybe Text

EC2-VPC

The ID that AWS assigns to represent the allocation of the Elastic IP address for use with a VPC.

AssignPrivateIpAddresses

data AssignPrivateIpAddresses

Assigns one or more secondary private IP addresses to the specified network interface.

You can specify one or more specific secondary IP addresses, or you can specify the number of secondary IP addresses to be automatically assigned within the subnet's CIDR block range.

The number of secondary IP addresses that you can assign to an instance varies by instance type.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-AssignPrivateIpAddresses.html

Constructors

AssignPrivateIpAddresses

Fields

apiaNetworkInterfaceId :: !Text

The ID of the network interface.

apiaPrivateIpAddress :: [Text]

The IP addresses to be assigned as a secondary private IP address to the network interface.

apiaSecondaryPrivateIpAddressCount :: Maybe Integer

The number of secondary IP addresses to assign to the network interface.

apiaAllowReassignment :: Maybe Bool

Specifies whether to allow an IP address that is already assigned to another network interface or instance to be reassigned to the specified network interface.

data AssignPrivateIpAddressesResponse

Constructors

AssignPrivateIpAddressesResponse

Fields

apiaRequestId :: !Text

The ID of the request.

apiaReturn :: !Bool

Returns true if the request succeeds. Otherwise, returns an error.

AssociateAddress

data AssociateAddress

Associates an Elastic IP address with an instance or a network interface.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-AssociateAddress.html

Constructors

AssociateAddress

Fields

abPublicIp :: !Text

The Elastic IP address.

abInstanceId :: !Text

The ID of the instance. The operation fails if you specify an instance ID unless exactly one network interface is attached.

abAllocationId :: Maybe Text

EC2-VPC

The allocation ID.

abNetworkInterfaceId :: Maybe Text

EC2-VPC

The ID of the network interface.

abPrivateIpAddress :: Maybe Text

EC2-VPC

The primary or secondary private IP address to associate with the Elastic IP address. If no private IP address is specified, the Elastic IP address is associated with the primary private IP address.

abAllowReassociation :: Maybe Bool

EC2-VPC

Allows an Elastic IP address that is already associated with an instance or network interface to be re-associated with the specified instance or network interface. Otherwise, the operation fails.

data AssociateAddressResponse

Constructors

AssociateAddressResponse

Fields

abrRequestId :: !Text

The ID of the request.

abrReturn :: !Bool

Returns true if the request succeeds. Otherwise, returns an error.

abrAssociationId :: Maybe Text

EC2-VPC

The ID that represents the association of the Elastic IP address with an instance.

AssociateDhcpOptions

data AssociateDhcpOptions

Associates a set of DHCP options (that you've previously created) with the specified VPC, or associates no DHCP options with the VPC.

After you associate the options with the VPC, any existing instances and all new instances that you launch in that VPC use the options. You don't need to restart or relaunch the instances. They automatically pick up the changes within a few hours, depending on how frequently the instance renews its DHCP lease. You can explicitly renew the lease using the operating system on the instance.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-AssociateDhcpOptions.html

Constructors

AssociateDhcpOptions

Fields

adoDhcpOptionsId :: !Text

The ID of the DHCP options set, or default to associate no DHCP options with the VPC.

adoVpcId :: !Text

The ID of the VPC.

data AssociateDhcpOptionsResponse

Constructors

AssociateDhcpOptionsResponse

Fields

adorRequestId :: !Text

The ID of the request.

adorReturn :: !Bool

Returns true if the request succeeds. Otherwise, returns an error.

AssociateRouteTable

data AssociateRouteTable

Associates a subnet with a route table.

The subnet and route table must be in the same VPC. This association causes traffic originating from the subnet to be routed according to the routes in the route table. The action returns an association ID, which you need in order to disassociate the route table from the subnet later. A route table can be associated with multiple subnets.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-AssociateRouteTable.html

Constructors

AssociateRouteTable

Fields

artRouteTableId :: !Text

The ID of the route table.

artSubnetId :: !Text

The ID of the subnet.

data AssociateRouteTableResponse

Constructors

AssociateRouteTableResponse

Fields

artrRequestId :: !Text

The ID of the request.

artrAssociationId :: !Text

The route table association ID (needed to disassociate the route table).

AttachInternetGateway

data AttachInternetGateway

Attaches an Internet gateway to a VPC, enabling connectivity between the Internet and the VPC.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-AttachInternetGateway.html

Constructors

AttachInternetGateway

Fields

aigInternetGatewayId :: !Text

The ID of the Internet gateway.

aigVpcId :: !Text

The ID of the VPC.

data AttachInternetGatewayResponse

Constructors

AttachInternetGatewayResponse

Fields

aigrRequestId :: !Text

The ID of the request.

aigrReturn :: !Bool

Returns true if the request succeeds. Otherwise, returns an error.

AttachNetworkInterface

data AttachNetworkInterface

Attaches a network interface to an instance.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-AttachNetworkInterface.html

Constructors

AttachNetworkInterface

Fields

aniNetworkInterfaceId :: !Text

The ID of the network interface.

aniInstanceId :: !Text

The ID of the instance.

aniDeviceIndex :: !Integer

The index of the device for the network interface attachment.

data AttachNetworkInterfaceResponse

Constructors

AttachNetworkInterfaceResponse

Fields

anirRequestId :: !Text

The ID of the attachment request.

anirAttachmentId :: !Text

The ID of the network interface attachment.

AttachVolume

data AttachVolume

Attaches an Amazon EBS volume to a running or stopped instance and exposes it to the instance with the specified device name.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-AttachVolume.html

Constructors

AttachVolume

Fields

avVolumeId :: !Text

The ID of the Amazon EBS volume. The volume and instance must be within the same Availability Zone.

avInstanceId :: !Text

The ID of the instance.

avDevice :: !Text

The device name to expose to the instance (for example, devsdh or xvdh).

data AttachVolumeResponse

Constructors

AttachVolumeResponse

Fields

avrRequestId :: !Text

The ID of the request.

avrVolumeId :: !Text

The ID of the volume.

avrInstanceId :: !Text

The ID of the instance.

avrDevice :: !Text

The device name.

avrStatus :: !VolumeStatus

The attachment state of the volume.

avrAttachTime :: !UTCTime

The time stamp when the attachment initiated.

AttachVpnGateway

data AttachVpnGateway

Attaches a virtual private gateway to a VPC.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-AttachVpnGateway.html

Constructors

AttachVpnGateway

Fields

avgVpnGatewayId :: !Text

The ID of the virtual private gateway.

avgVpcId :: !Text

The ID of the VPC.

data AttachVpnGatewayResponse

Constructors

AttachVpnGatewayResponse

Fields

avgrRequestId :: !Text

The ID of the request.

avgrAttachment :: !Attachment

Information about the attachment.

AuthorizeSecurityGroupEgress

data AuthorizeSecurityGroupEgress

Adds one or more egress rules to a security group for use with a VPC.

Specifically, this action permits instances to send traffic to one or more destination CIDR IP address ranges, or to one or more destination security groups for the same VPC.

This action doesn't apply to security groups for use in EC2-Classic.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-AuthorizeSecurityGroupEgress.html

Constructors

AuthorizeSecurityGroupEgress

Fields

asgeGroupId :: !Text

The ID of the security group to modify.

asgeIpPermissions :: [IpPermissionType]

The IP protocol name or number (see Protocol Numbers).

data AuthorizeSecurityGroupEgressResponse

Constructors

AuthorizeSecurityGroupEgressResponse

Fields

asgerRequestId :: !Text

The ID of the request.

asgerReturn :: !Bool

Returns true if the request succeeds. Otherwise, returns an error.

AuthorizeSecurityGroupIngress

data AuthorizeSecurityGroupIngress

Adds one or more ingress rules to a security group.

Rule changes are propagated to instances within the security group as quickly as possible.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-AuthorizeSecurityGroupIngress.html

Constructors

AuthorizeSecurityGroupIngress

Fields

asgiGroupId :: Maybe Text

The ID of the security group to modify. The security group must belong to your account.

asgiGroupName :: Maybe Text

The name of the security group to modify.

asgiIpPermissions :: [IpPermissionType]

The IP protocol name or number (see Protocol Numbers). For EC2-Classic, security groups can have rules only for TCP, UDP, and ICMP. For EC2-VPC, security groups can have rules assigned to any protocol number.

data AuthorizeSecurityGroupIngressResponse

Constructors

AuthorizeSecurityGroupIngressResponse

Fields

asgirRequestId :: !Text

The ID of the request.

asgirReturn :: !Bool

Returns true if the request succeeds. Otherwise, returns an error.

BundleInstance

data BundleInstance

Bundles an Amazon instance store-backed Windows instance.During bundling, only the root device volume (C:) is bundled. Data on other instance store volumes is not preserved.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-BundleInstance.html

Constructors

BundleInstance

Fields

biInstanceId :: !Text

The ID of the instance to bundle.

biStorage :: !BundleInstanceTaskStorage

Bundle storage instructions.

data BundleInstanceResponse

Constructors

BundleInstanceResponse

Fields

birRequestId :: !Text

The ID of the request.

birBundleInstanceTask :: !BundleInstanceTask

The bundle task.

CancelBundleTask

data CancelBundleTask

Cancels a bundling operation for an instance store-backed Windows instance.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-CancelBundleTask.html

Constructors

CancelBundleTask

Fields

cbtBundleId :: !Text

The ID of the bundle task.

data CancelBundleTaskResponse

Constructors

CancelBundleTaskResponse

Fields

cbtRequestId :: !Text

The ID of the request.

cbtBundleInstanceTask :: !BundleInstanceTask

The bundle task.

CancelConversionTask

data CancelConversionTask

Cancels an active conversion task.

The task can be the import of an instance or volume. The action removes all artifacts of the conversion, including a partially uploaded volume or instance. If the conversion is complete or is in the process of transferring the final disk image, the command fails and returns an exception.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-CancelConversionTask.html

Constructors

CancelConversionTask

Fields

cctConversionTaskId :: !Text

The ID of the conversion task.

data CancelConversionTaskResponse

Constructors

CancelConversionTaskResponse

Fields

cctRequestId :: !Text

The ID of the request.

cctReturn :: !Bool

Returns true if the request succeeds. Otherwise, returns an error.

CancelExportTask

data CancelExportTask

Cancels an active export task. The request removes all artifacts of the export, including any partially created Amazon S3 objects. If the export task is complete or is in the process of transferring the final disk image, the command fails and returns an error.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-CancelExportTask.html

Constructors

CancelExportTask

Fields

cetExportTaskId :: !Text

The ID of the export task. This is the ID returned by CreateInstanceExportTask.

data CancelExportTaskResponse

Constructors

CancelExportTaskResponse

Fields

cetRequestId :: !Text

The ID of the request.

cetReturn :: !Bool

Returns true if the request succeeds. Otherwise, returns an error.

CancelReservedInstancesListing

data CancelReservedInstancesListing

Cancels the specified Reserved Instance listing in the Reserved Instance Marketplace.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-CancelReservedInstancesListing.html

Constructors

CancelReservedInstancesListing

Fields

crilReservedInstancesListingId :: !Text

The ID of the Reserved Instance listing to be canceled.

data CancelReservedInstancesListingResponse

Constructors

CancelReservedInstancesListingResponse

Fields

crilRequestId :: !Text

The ID of the request.

crilReservedInstancesListingsSet :: !DescribeReservedInstancesListingsResponseSetItemType

The Reserved Instance listing for cancellation.

CancelSpotInstanceRequests

data CancelSpotInstanceRequests

Cancels one or more Spot Instance requests.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-CancelSpotInstanceRequests.html

Constructors

CancelSpotInstanceRequests

Fields

csirSpotInstanceRequestId :: [Text]

One or more Spot Instance request IDs.

data CancelSpotInstanceRequestsResponse

Constructors

CancelSpotInstanceRequestsResponse

Fields

csirRequestId :: !Text

The ID of the request.

csirSpotInstanceRequestSet :: [CancelSpotInstanceRequestsResponseSetItemType]

A list of Spot Instance requests. Each request is wrapped in an item element.

CreateKeyPair

data CreateKeyPair

Creates a 2048-bit RSA key pair with the specified name. Amazon EC2 stores the public key and displays the private key for you to save to a file. The private key is returned as an unencrypted PEM encoded PKCS#8 private key. If a key with the specified name already exists, Amazon EC2 returns an error. You can have up to five thousand key pairs per region.For more information about key pairs, see Key Pairs in the Amazon Elastic Compute Cloud User Guide.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-CreateKeyPair.html

Constructors

CreateKeyPair

Fields

ckpKeyName :: !Text

A unique name for the key pair.

data CreateKeyPairResponse

Constructors

CreateKeyPairResponse

Fields

ckpRequestId :: !Text

The ID of the request.

ckqKeyName :: !Text

The name of the key pair name.

ckqKeyFingerprint :: !Text

A SHA-1 digest of the DER encoded private key.

ckqKeyMaterial :: !Text

An unencrypted PEM encoded RSA private key.

CreateSecurityGroup

data CreateSecurityGroup

Creates a security group.

A security group is for use with instances either in the EC2-Classic platform or in a specific VPC.

When you create a security group, you specify a friendly name of your choice. You can have a security group for use in EC2-Classic with the same name as a security group for use in a VPC. However, you can't have two security groups for use in EC2-Classic with the same name or two security groups for use in a VPC with the same name.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-CreateSecurityGroup.html

Constructors

CreateSecurityGroup

Fields

csgGroupName :: !Text

The name of the security group.

csgGroupDescription :: !Text

A description for the security group. This is informational only.

csgVpcId :: Maybe Text

EC2-VPC

The ID of the VPC.

data CreateSecurityGroupResponse

Constructors

CreateSecurityGroupResponse

Fields

csgrRequestId :: !Text

The ID of the request.

csgrReturn :: !Bool

Returns true if the request succeeds. Otherwise, returns an error.

csgrGroupId :: !Text

The ID of the new security group.

CreateTags

data CreateTags

Adds or overwrites one or more tags for the specified EC2 resource or resources. Each resource can have a maximum of 10 tags. Each tag consists of a key and optional value. Tag keys must be unique per resource.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-CreateTags.html

Constructors

CreateTags

Fields

ctResourceId :: [Text]

The IDs of one or more resources to tag. For example, ami-1a2b3c4d.

ctTag :: [ResourceTagSetItemType]

The key for a tag.

data CreateTagsResponse

Constructors

CreateTagsResponse

Fields

cvRequestId :: !Text

The ID of the request.

cvReturn :: !Bool

Returns true if the request succeeds. Otherwise, returns an error.

DeleteSecurityGroup

data DeleteSecurityGroup

Deletes a security group.A security group is for use with instances either in the EC2-Classic platform or in a specific VPC.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-DeleteSecurityGroup.html

Constructors

DeleteSecurityGroup

Fields

dsgGroupName :: Maybe Text

The name of the security group.

dsgGroupId :: Maybe Text

The ID of the security group.

data DeleteSecurityGroupResponse

Constructors

DeleteSecurityGroupResponse

Fields

dsgrRequestId :: !Text

The ID of the request.

dsgrReturn :: !Bool

Returns true if the request succeeds. Otherwise, returns an error.

DescribeAvailabilityZones

data DescribeAvailabilityZones

Describes one or more of the Availability Zones that are available to you. The results include zones only for the region you're currently using.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-DescribeAvailabilityZones.html

Constructors

DescribeAvailabilityZones

Fields

dazZoneName :: [Text]

One or more Availability Zone names.

dazFilter :: [Filter]

The name of a filter.

data DescribeAvailabilityZonesResponse

Constructors

DescribeAvailabilityZonesResponse

Fields

dazrRequestId :: !Text

The ID of the request.

dazrAvailabilityZoneInfo :: [AvailabilityZoneItemType]

A list of Availability Zones.

DescribeImages

data DescribeImages

Describes one or more of the images (AMIs, AKIs, and ARIs) available to you. Images available to you include public images, private images that you own, and private images owned by other AWS accounts but for which you have explicit launch permissions.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-DescribeImages.html

Constructors

DescribeImages

Fields

djExecutableBy :: [Text]

Describes the images for which the specified user has explicit launch permissions. The user ID can be an AWS account ID, self to return images for which the sender of the request has explicit launch permissions, or all to return AMIs with public launch permissions.

djImageId :: [Text]

One or more image IDs.

djOwner :: [Text]

Describes images owned by the specified owners. Use the IDs amazon, aws-marketplace, and self to describe images owned by Amazon, AWS Marketplace, or you, respectively.

djFilter :: [Filter]

The name of a filter.

data DescribeImagesResponse

Constructors

DescribeImagesResponse

Fields

djRequestId :: !Text

The ID of the request.

djImagesSet :: [DescribeImagesResponseItemType]

A list of images.

DescribeInstances

data DescribeInstances

Describes one or more of your instances.

If you specify one or more instance IDs, Amazon EC2 returns information for those instances.

If you do not specify instance IDs, Amazon EC2 returns information for all relevant instances.

If you specify an invalid instance ID, an error is returned.

If you specify an instance that you do not own, it is not included in the returned results.

Recently terminated instances might appear in the returned results. This interval is usually less than one hour.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-DescribeInstances.html

Constructors

DescribeInstances

Fields

diInstanceId :: [Text]

One or more instance IDs.

diFilter :: [Filter]

The name of a filter.

data DescribeInstancesResponse

Constructors

DescribeInstancesResponse

Fields

dirRequestId :: !Text

The ID of the request.

dirReservationSet :: [ReservationInfoType]

A list of reservations, each one wrapped in an item element.

DescribeKeyPairs

data DescribeKeyPairs

Describes one or more of your key pairs.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-DescribeKeyPairs.html

Constructors

DescribeKeyPairs

Fields

dkqKeyName :: [Text]

One or more key pair names.

dkqFilter :: [Filter]

The name of a filter.

data DescribeKeyPairsResponse

Constructors

DescribeKeyPairsResponse

Fields

dkqRequestId :: !Text

The ID of the request.

dkqKeySet :: [DescribeKeyPairsResponseItemType]

A list of key pairs.

DescribeRegions

data DescribeRegions

Describes one or more regions that are currently available to you.For a list of the regions supported by Amazon EC2, see Regions and Endpoints.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-DescribeRegions.html

Constructors

DescribeRegions

Fields

drRegionName :: [Region]

One or more region names.

drFilter :: [Filter]

The name of a filter.

data DescribeRegionsResponse

Constructors

DescribeRegionsResponse

Fields

drrRequestId :: !Text

The ID of the request.

drrRegionInfo :: [RegionItemType]

A list of regions, each one wrapped in an item element.

DescribeSecurityGroups

data DescribeSecurityGroups

A security group is for use with instances either in the EC2-Classic platform or in a specific VPC.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-DescribeSecurityGroups.html

Constructors

DescribeSecurityGroups

Fields

dshGroupName :: [Text]

A list of group names.

dshGroupId :: [Text]

A list of group ids.

dshFilter :: [Filter]

The name of a filter.

data DescribeSecurityGroupsResponse

Constructors

DescribeSecurityGroupsResponse

Fields

dshrRequestId :: !Text

The ID of the request.

dshrSecurityGroupInfo :: [SecurityGroupItemType]

A list of security groups.

DescribeTags

data DescribeTags

Describes one or more of the tags for your EC2 resources.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-DescribeTags.html

Constructors

DescribeTags

Fields

dtagsFilter :: [TagFilter]

The name of a filter.

data DescribeTagsResponse

Constructors

DescribeTagsResponse

Fields

dtagsrRequestId :: !Text

The ID of the request.

dtagsrTagSet :: [TagSetItemType]

A list of tags.

RevokeSecurityGroupEgress

data RevokeSecurityGroupEgress

Removes one or more egress rules from a security group for EC2-VPC.

The values that you specify in the revoke request (for example, ports) must match the existing rule's values for the rule to be revoked.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-RevokeSecurityGroupEgress.html

Constructors

RevokeSecurityGroupEgress

Fields

rsgeGroupId :: !Text

The ID of the security group to modify.

rsgeIpPermissions :: [IpPermissionType]

The IP protocol name or number (see Protocol Numbers).

data RevokeSecurityGroupEgressResponse

Constructors

RevokeSecurityGroupEgressResponse

Fields

rsgerRequestId :: !Text

The ID of the request.

rsgerReturn :: !Bool

Returns true if the request succeeds. Otherwise, returns an error.

RevokeSecurityGroupIngress

data RevokeSecurityGroupIngress

Removes one or more ingress rules from a security group.

The values that you specify in the revoke request (for example, ports) must match the existing rule's values for the rule to be removed.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-RevokeSecurityGroupIngress.html

Constructors

RevokeSecurityGroupIngress

Fields

rsgiGroupId :: Maybe Text

The ID of the security group to modify. The security group must belong to your account.

rsgiGroupName :: Maybe Text

The name of the security group to modify.

rsgiIpPermissions :: [IpPermissionType]

The IP protocol name or number (see Protocol Numbers). For EC2-Classic, security groups can have rules only for TCP, UDP, and ICMP. For EC2-VPC, security groups can have rules assigned to any protocol number.

data RevokeSecurityGroupIngressResponse

Constructors

RevokeSecurityGroupIngressResponse

Fields

rsgirRequestId :: !Text

The ID of the request.

rsgirReturn :: !Bool

Returns true if the request succeeds. Otherwise, returns an error.

RunInstances

data RunInstances

Launches the specified number of instances of an AMI for which you have permissions.When you launch an instance, it enters the pending state. After the instance is ready for you, it enters the running state. To check the state of your instance, call DescribeInstances.If you don't specify a security group when launching an instance, Amazon EC2 uses the default security group.Linux instances have access to the public key of the key pair at boot. You can use this key to provide secure access to the instance. Amazon EC2 public images use this feature to provide secure access without passwords.You can provide optional user data when launching an instance.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-RunInstances.html

Constructors

RunInstances

Fields

riImageId :: !Text

The ID of the AMI, which you can get by calling DescribeImages.

riMinCount :: !Integer

The minimum number of instances to launch. If you specify a minimum that is more instances than Amazon EC2 can launch in the target Availability Zone, Amazon EC2 launches no instances.

riMaxCount :: !Integer

The maximum number of instances to launch. If you specify more instances than Amazon EC2 can launch in the target Availability Zone, Amazon EC2 launches the largest possible number of instances above MinCount.

riKeyName :: Maybe Text

The name of the key pair. You can create a key pair using CreateKeyPair or ImportKeyPair.

riSecurityGroupId :: [Text]

One or more security group IDs. You can create a security group using CreateSecurityGroup.

riSecurityGroup :: [Text]

EC2-Classic, default VPC

One or more security group names. For a nondefault VPC, you must use SecurityGroupId.

riUserData :: Maybe Text

The Base64-encoded MIME user data for the instances.

riInstanceType :: Maybe InstanceType

The instance type. Defaults to m1.small if absent.

riPlacement :: Maybe PlacementType

The Availability Zone for the instance.

rjKernelId :: Maybe Text

The ID of the kernel.

rjRamdiskId :: Maybe Text

The ID of the RAM disk.

rjBlockDeviceMapping :: [InstanceBlockDeviceMappingItemType]

The device name exposed to the instance (for example, devsdh or xvdh).

rjMonitoring :: Maybe MonitoringInstanceType

Enables monitoring for the instance.

rjSubnetId :: Maybe Text

EC2-VPC

The ID of the subnet to launch the instance into.

rjDisableApiTermination :: Maybe Bool

If you set this parameter to true, you can't terminate the instance using the Amazon EC2 console, CLI, or API; otherwise, you can. If you set this parameter to true and then later want to be able to terminate the instance, you must first change the value of the disableApiTermination attribute to false using ModifyInstanceAttribute. Alternatively, if you set InstanceInitiatedShutdownBehavior to terminate, you can terminate the instance by running the shutdown command from the instance.

rjInstanceInitiatedShutdownBehavior :: Maybe Text

Indicates whether an instance stops or terminates when you initiate shutdown from the instance (using the operating system command for system shutdown).

rjPrivateIpAddress :: Maybe Text

EC2-VPC

The primary IP address. You must specify a value from the IP address range of the subnet.

rjClientToken :: Maybe Text

Unique, case-sensitive identifier you provide to ensure idempotency of the request.

rjNetworkInterface :: [NetworkInterfaceType]

An existing interface to attach to a single instance. Requires n=1 instances.

rjIamInstanceProfile :: [IamInstanceProfileRequestType]

The Amazon Resource Name (ARN) of the IAM instance profile to associate with the instances.

rjEbsOptimized :: Maybe Bool

Indicates whether the instance is optimized for EBS I/O. This optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal Amazon EBS I/O performance. This optimization isn't available with all instance types. Additional usage charges apply when using an EBS-optimized instance.

data RunInstancesResponse

Constructors

RunInstancesResponse

Fields

rirRequestId :: !Text

The ID of the request.

rirReservationId :: !Text

The ID of the reservation.

rirOwnerId :: !Text

The ID of the AWS account that owns the reservation.

rirGroupSet :: [GroupItemType]

A list of security groups the instance belongs to. Each group is wrapped in an item element.

rirInstancesSet :: [RunningInstancesItemType]

A list of instances. Each instance is wrapped in an item element.

rirRequesterId :: Maybe Text

The ID of the requester that launched the instances on your behalf (for example, AWS Management Console, Auto Scaling).

Data Types

ec2 :: Service

Currently supported version (2013-07-15) of the EC2 service.

data EC2Error

Constructors

EC2Error

Fields

ecCode :: !Text

ecMessage :: !Text

data EC2ErrorResponse

Constructors

EC2ErrorResponse

Fields

eerErrors :: [EC2Error]

eerRequestID :: !Text

data Protocol

Constructors

TCP

UDP

ICMP

data AddressDomain

Constructors

AddressStandard

AddressVPC

data VolumeStatus

Constructors

Attaching

Attached

Detaching

Detached

data Attachment

Constructors

Attachment

Fields

atVpcId :: !Text

The ID of the VPC.

atState :: !Text

The current state of the attachment.

data AvailabilityZoneItemType

Constructors

AvailabilityZoneItemType

Fields

azitZoneName :: !AvailabilityZone

The name of the Availability Zone.

azitZoneState :: !Text

The state of the Availability Zone. FIXME: Should be - available | impaired | unavailable

azitRegionName :: !Region

The name of the region.

azitMessageSet :: [AvailabilityZoneMessageType]

Any messages about the Availability Zone.

data AvailabilityZoneMessageType

Constructors

AvailabilityZoneMessageType

Fields

azmtMessage :: !Text

The message about the Availability Zone.

data BlockDeviceMappingItemType

Constructors

BlockDeviceMappingItemType

Fields

bdmitDeviceName :: !Text

The device name exposed to the instance (for example, devsdh).

bdmitVirtualName :: Maybe Text

The virtual device name.

bdmitEbs :: Maybe EbsBlockDeviceType

Parameters used to automatically set up Amazon EBS volumes when the instance is launched.

bdmitNoDevice :: Maybe Text

Include this empty element to suppress the specified device included in the block device mapping of the AMI.

data BundleInstanceS3Storage

Constructors

BundleInstanceS3Storage

Fields

bissAwsAccessKeyId :: !Text

The access key ID of the owner of the bucket.

bissBucket :: !Text

The bucket in which to store the AMI. You can specify a bucket that you already own or a new bucket that Amazon EC2 creates on your behalf. If you specify a bucket that belongs to someone else, Amazon EC2 returns an error.

bissPrefix :: !Text

The beginning of the file name of the AMI.

bissUploadPolicy :: !Text

A Base64-encoded Amazon S3 upload policy that gives Amazon EC2 permission to upload items into Amazon S3 on the user's behalf.

bissUploadPolicySignature :: !Text

The signature of the Base64 encoded JSON document.

data BundleInstanceTaskStorage

Constructors

BundleInstanceTaskStorage

Fields

bitsS3 :: !BundleInstanceS3Storage

An Amazon S3 storage location.

data BundleInstanceTaskError

Constructors

BundleInstanceTaskError

Fields

biteCode :: !Text

The error code.

biteMessage :: !Text

The error message.

data BundleInstanceState

Constructors

Pending

WaitingForShutdown

Bundling

Storing

Cancelling

Complete

Failed

data BundleInstanceTask

Constructors

BundleInstanceTask

Fields

bitInstanceId :: !Text

The ID of the instance associated with this bundle task.

bitBundleId :: !Text

The ID for this bundle task.

bitState :: !BundleInstanceState

The state of the task.

bitStartTime :: !UTCTime

The time this task started.

bitUpdateTime :: !UTCTime

The time of the most recent update for the task.

bitStorage :: !BundleInstanceTaskStorage

The Amazon S3 storage locations.

bitProgress :: !Text

The level of task completion, as a percent (for example, 20%).

bitError :: Maybe BundleInstanceTaskError

If the task fails, a description of the error.

data CancelSpotInstanceRequestsResponseSetItemType

Constructors

CancelSpotInstanceRequestsResponseSetItemType

Fields

csirrsitSpotInstanceRequestId :: !Text

The ID of the Spot Instance request.

csirrsitState :: !Text

The state of the Spot Instance request.

data DescribeImagesResponseItemType

Constructors

DescribeImagesResponseItemType

Fields

diritImageId :: !Text

The ID of the AMI.

diritImageLocation :: !Text

The location of the AMI.

diritImageState :: !Text

Current state of the AMI. If the operation returns available, the image is successfully registered and available for launching.

diritImageOwnerId :: !Text

AWS account ID of the image owner.

diritIsPublic :: !Bool

Indicates whether the image has public launch permissions. The value is true if this image has public launch permissions or false if it has only implicit and explicit launch permissions.

diritProductCodes :: [ProductCodesSetItemType]

Any product codes associated with the AMI.

diritArchitecture :: !Text

The architecture of the image.

diritImageType :: !Text

The type of image.

diritKernelId :: Maybe Text

The kernel associated with the image, if any. Only applicable for machine images.

diritRamdiskId :: Maybe Text

The RAM disk associated with the image, if any. Only applicable for machine images.

diritPlatform :: Maybe Text

The value is Windows for Windows AMIs; otherwise blank.

diritStateReason :: Maybe StateReasonType

The reason for the state change.

diritImageOwnerAlias :: Maybe Text

The AWS account alias (for example, amazon, self, etc.) or AWS account ID that owns the AMI.

diritName :: !Text

The name of the AMI that was provided during image creation.

diritDescription :: Maybe Text

The description of the AMI that was provided during image creation.

diritRootDeviceType :: !Text

The type of root device used by the AMI. The AMI can use an Amazon EBS volume or an instance store volume.

diritRootDeviceName :: !Text

The device name of the root device (for example, devsda1 or xvda).

diritBlockDeviceMapping :: [BlockDeviceMappingItemType]

Any block device mapping entries.

diritVirtualizationType :: !Text

The type of virtualization of the AMI.

diritTagSet :: [ResourceTagSetItemType]

Any tags assigned to the resource.

diritHypervisor :: !Text

The image's hypervisor type.

data DescribeKeyPairsResponseItemType

Constructors

DescribeKeyPairsResponseItemType

Fields

dkpritKeyName :: !Text

The name of the key pair.

dkpritKeyFingerprint :: !Text

If you used CreateKeyPair to create the key pair, this is the SHA-1 digest of the DER encoded private key. If you used ImportKeyPair to provide AWS the public key, this is the MD5 public key fingerprint as specified in section 4 of RFC4716.

data DescribeReservedInstancesListingsResponseSetItemType

Constructors

DescribeReservedInstancesListingsResponseSetItemType

Fields

drilrsitReservedInstancesListingId :: !Text

The ID of the Reserved Instance listing.

drilrsitReservedInstancesId :: !Text

The ID of the Reserved Instance.

drilrsitCreateDate :: !UTCTime

The time the listing was created.

drilrsitUpdateDate :: !UTCTime

The last modified timestamp of the listing.

drilrsitStatus :: !Text

The status of the Reserved Instance listing.

drilrsitStatusMessage :: !Text

The reason for the current status of the Reserved Instance listing. The response can be blank.

drilrsitInstanceCounts :: [InstanceCountsSetItemType]

The number of instances in this state.

drilrsitPriceSchedules :: [PriceScheduleSetItemType]

The price of the Reserved Instance listing.

drilrsitTagSet :: [ResourceTagSetItemType]

The tags assigned to the resource. Each tag's information is wrapped in an item element.

drilrsitClientToken :: !Text

The idempotency token you provided when you created the listing.

data EbsBlockDeviceType

Constructors

EbsBlockDeviceType

Fields

ebdtSnapshotId :: !Text

The ID of the snapshot.

ebdtVolumeSize :: !Integer

The size of the volume, in GiB.

ebdtDeleteOnTermination :: !Bool

Indicates whether the Amazon EBS volume is deleted on instance termination.

ebdtVolumeType :: !Text

The volume type.

ebdtIops :: Maybe Integer

The number of I/O operations per second (IOPS) that the volume supports.

data EbsInstanceBlockDeviceMappingResponseType

Constructors

EbsInstanceBlockDeviceMappingResponseType

Fields

eibdmrtVolumeId :: !Text

The ID of the Amazon EBS volume.

eibdmrtStatus :: !Text

The attachment state.

eibdmrtAttachTime :: !UTCTime

The time stamp when the attachment initiated.

eibdmrtDeleteOnTermination :: !Bool

Indicates whether the volume is deleted on instance termination.

data GroupItemType

Constructors

GroupItemType

Fields

gitGroupId :: !Text

The ID of the security group.

gitGroupName :: !Text

The name of the security group.

data IamInstanceProfileRequestType

Constructors

IamInstanceProfileRequestType

Fields

iiprtArn :: Maybe Text

The Amazon Resource Name (ARN) of the instance profile.

iiprtName :: Maybe Text

The name of the instance profile.

data IamInstanceProfileResponseType

Constructors

IamInstanceProfileResponseType

Fields

iipruArn :: !Text

The Amazon Resource Name (ARN) of the instance profile.

iipruId :: !Text

The ID of the instance profile.

data InstanceBlockDeviceMappingItemType

Constructors

InstanceBlockDeviceMappingItemType

Fields

ibdmitDeviceName :: !Text

The device name exposed to the instance (for example, devsdh or xvdh).

ibdmitVirtualName :: !Text

The virtual device name.

ibdmitEbs :: !InstanceEbsBlockDeviceType

Parameters used to automatically set up Amazon EBS volumes when the instance is launched.

ibdmitNoDevice :: !Text

Include this empty element to suppress the specified device included in the block device mapping of the AMI.

data InstanceBlockDeviceMappingResponseItemType

Constructors

InstanceBlockDeviceMappingResponseItemType

Fields

ibdmritDeviceName :: !Text

The device name exposed to the instance (for example, devsdh, or xvdh).

ibdmritEbs :: !EbsInstanceBlockDeviceMappingResponseType

Parameters used to automatically set up Amazon EBS volumes when the instance is launched.

data InstanceCountsSetItemType

Constructors

InstanceCountsSetItemType

Fields

icsitState :: !Text

The states of the listed Reserved Instances.

icsitInstanceCount :: !Integer

The number of listed Reserved Instances in the state specified by the state.

data InstanceEbsBlockDeviceType

Constructors

InstanceEbsBlockDeviceType

Fields

iebdtDeleteOnTermination :: !Bool

Indicates whether the volume is deleted on instance termination.

iebdtVolumeId :: !Text

The ID of the volume.

data InstanceMonitoringStateType

Constructors

InstanceMonitoringStateType

Fields

imstState :: !Text

The state of monitoring for the instance.

data InstanceNetworkInterfaceAssociationType

Constructors

InstanceNetworkInterfaceAssociationType

Fields

iniatPublicIp :: !Text

The address of the Elastic IP address bound to the network interface.

iniatPublicDnsName :: !Text

The public DNS name.

iniatIpOwnerId :: !Text

The ID of the owner of the Elastic IP address.

data InstanceNetworkInterfaceAttachmentType

Constructors

InstanceNetworkInterfaceAttachmentType

Fields

iniatAttachmentID :: !Text

The ID of the network interface attachment.

iniatDeviceIndex :: !Integer

The index of the device on the instance for the network interface attachment.

iniatStatus :: !Text

The attachment state.

iniatAttachTime :: !UTCTime

The time stamp when the attachment initiated.

iniatDeleteOnTermination :: !Bool

Indicates whether the network interface is deleted when the instance is terminated.

data InstanceNetworkInterfaceSetItemType

Constructors

InstanceNetworkInterfaceSetItemType

Fields

inisitNetworkInterfaceId :: !Text

The ID of the network interface.

inisitSubnetId :: !Text

The ID of the subnet.

inisitVpcId :: !Text

The ID of the VPC.

inisitDescription :: !Text

The description.

inisitOwnerId :: !Text

The ID of the customer who created the network interface.

inisitStatus :: !Text

The status of the network interface.

inisitMacAddress :: !Text

The MAC address.

inisitPrivateIpAddress :: !Text

The IP address of the network interface within the subnet.

inisitPrivateDnsName :: !Text

The private DNS name.

inisitSourceDestCheck :: !Bool

Indicates whether to validate network traffic to or from this network interface.

inisitGroupSet :: [GroupItemType]

A security group.

inisitAttachment :: !InstanceNetworkInterfaceAttachmentType

The network interface attachment.

inisitAssociation :: !InstanceNetworkInterfaceAssociationType

The association information for an Elastic IP associated with the network interface.

inisitPrivateIpAddressesSet :: [InstancePrivateIpAddressesSetItemType]

The private IP addresses associated with the network interface.

data InstancePrivateIpAddressesSetItemType

Constructors

InstancePrivateIpAddressesSetItemType

Fields

ipiasitPrivateIpAddress :: !Text

The private IP address of the network interface

ipiasitPrivateDnsName :: !Text

The private DNS name.

ipiasitPrimary :: !Bool

Indicates whether this IP address is the primary private IP address of the network interface.

ipiasitAssociation :: !InstanceNetworkInterfaceAssociationType

The association information for an Elastic IP address associated with the network interface.

data InstanceStateType

Constructors

InstanceStateType

Fields

istCode :: !Integer

The low byte represents the state. The high byte is an opaque internal value and should be ignored.

istName :: !Text

The current state of the instance.

data UserIdGroupPair

Constructors

UserIdGroupPair

Fields

uigUserId :: Maybe Text

The ID of an AWS account. Cannot be used when specifying a CIDR IP address range.

uigGroupId :: Maybe Text

The ID of the security group in the specified AWS account. Cannot be used when specifying a CIDR IP address range.

uigGroupName :: Maybe Text

The name of the security group in the specified AWS account. Cannot be used when specifying a CIDR IP address range.

data IpPermissionType

Constructors

IpPermissionType

Fields

iptIpProtocol :: !Protocol

The protocol.

iptFromPort :: !Integer

The start of port range for the ICMP and UDP protocols, or an ICMP type number. A value of -1 indicates all ICMP types.

iptToPort :: !Integer

The end of port range for the ICMP and UDP protocols, or an ICMP code. A value of -1 indicates all ICMP codes for the given ICMP type.

iptGroups :: [UserIdGroupPair]

A list of security group and AWS account ID pairs.

iptIpRanges :: [IpRange]

A list of IP ranges.

data IpRange

Constructors

IpRange

Fields

irCidrIp :: !Text

The CIDR range. You can either specify a CIDR range or a source security group, not both.

data MonitoringInstanceType

Constructors

MonitoringInstanceType

Fields

mitEnabled :: !Bool

Indicates whether monitoring is enabled for the instance.

data NetworkInterfaceAssociationType

Constructors

NetworkInterfaceAssociationType

Fields

niatPublicIp :: !Text

The address of the Elastic IP address bound to the network interface.

niatPublicDnsName :: !Text

The public DNS name.

niatIpOwnerId :: !Text

The ID of the Elastic IP address owner.

niatAllocationID :: !Text

The allocation ID.

niatAssociationID :: !Text

The association ID.

data NetworkInterfaceAttachmentType

Constructors

NetworkInterfaceAttachmentType

Fields

niatAttachmentID :: !Text

The ID of the network interface attachment.

niatInstanceID :: !Text

The ID of the instance.

data NetworkInterfacePrivateIpAddressesSetItemType

Constructors

NetworkInterfacePrivateIpAddressesSetItemType

Fields

nipiasitPrivateIpAddress :: !Text

The private IP address of the network interface.

nipiasitPrivateDnsName :: !Text

The private DNS name.

nipiasitPrimary :: !Bool

Indicates whether this IP address is the primary private IP address of the network interface.

nipiasitAssociation :: !NetworkInterfaceAssociationType

The association information for an Elastic IP address associated with the network interface.

data NetworkInterfaceType

Constructors

NetworkInterfaceType

Fields

nitNetworkInterfaceId :: !Text

The ID of the network interface.

nitSubnetId :: !Text

The ID of the subnet.

niuNetworkInterfaceId :: !Text

The ID of the network interface.

niuSubnetId :: !Text

The ID of the subnet.

niuVpcId :: !Text

The ID of the VPC.

niuAvailabilityZone :: !Text

The Availability Zone.

niuDescription :: !Text

A description.

niuOwnerId :: !Text

The ID of the customer who created the interface.

niuRequesterId :: !Text

The ID of the entity that launched the instance on your behalf (for example, AWS Management Console or Auto Scaling)

niuRequesterManaged :: !Text

Indicates whether the network interface is being managed by AWS.

niuStatus :: !Text

The status of the network interface.

niuMacAddress :: !Text

The MAC address.

niuPrivateIpAddress :: !Text

The IP address of the network interface within the subnet.

niuPrivateDnsName :: !Text

The private DNS name.

niuSourceDestCheck :: !Bool

Indicates whether traffic to or from the instance is validated.

niuGroupSet :: !GroupItemType

The security group.

niuAttachment :: !NetworkInterfaceAttachmentType

The network interface attachment.

niuAssociation :: !NetworkInterfaceAssociationType

The association information for an Elastic IP associated with the network interface.

niuTagSet :: !ResourceTagSetItemType

The tags assigned to the resource.

niuPrivateIpAddressesSet :: !NetworkInterfacePrivateIpAddressesSetItemType

The private IP addresses associated with the network interface. [are returned in a set.]

data PlacementType

Constructors

PlacementType

Fields

pruAvailabilityZone :: Maybe AvailabilityZone

The Availability Zone of the instance.

pruGroupName :: Maybe Text

The name of the placement group the instance is in (for cluster compute instances).

pruTenancy :: Maybe Text

The tenancy of the instance (if the instance is running within a -- VPC). FIXME: switch to enum default | dedicated

data PriceScheduleSetItemType

Constructors

PriceScheduleSetItemType

Fields

pssitTerm :: !Integer

The number of months remaining in the reservation. For example, 2 is the second to the last month before the capacity reservation expires.

pssitPrice :: !Double

The fixed price for the term.

pssitCurrencyCode :: !Text

The currency for transacting the Reserved Instance resale. At this time, the only supported currency is USD.

pssitActive :: !Bool

The current price schedule, as determined by the term remaining for the Reserved Instance in the listing.

data ProductCodesSetItemType

Constructors

ProductCodesSetItemType

Fields

pcsitProductCode :: !Text

The product code.

pcsitType :: !Text

The type of product code.

data RegionItemType

Constructors

RegionItemType

Fields

ritRegionName :: !Text

The name of the region.

ritRegionEndpoint :: !Text

The region service endpoint.

data ReservationInfoType

Constructors

ReservationInfoType

Fields

ritReservationId :: !Text

The ID of the reservation.

ritOwnerId :: !Text

The ID of the AWS account that owns the reservation.

ritGroupSet :: [GroupItemType]

A list of security groups.

ritInstancesSet :: [RunningInstancesItemType]

A list of instances.

ritRequesterId :: Maybe Text

The ID of the requester that launched the instances on your behalf (for example, AWS Management Console or Auto Scaling).

data ResourceTagSetItemType

Constructors

ResourceTagSetItemType

Fields

rtsitKey :: !Text

The tag key.

rtsitValue :: !Text

The tag value.

data RunningInstancesItemType

Constructors

RunningInstancesItemType

Fields

riitInstanceId :: !Text

The ID of the instance launched.

riitImageId :: !Text

The ID of the AMI used to launch the instance.

riitInstanceState :: !InstanceStateType

The current state of the instance.

riitPrivateDnsName :: Maybe Text

The private DNS name assigned to the instance. This DNS name can only be used inside the Amazon EC2 network. This element remains empty until the instance enters the running state.

riitDnsName :: Maybe Text

The public DNS name assigned to the instance. This element remains empty until the instance enters the running state.

riitReason :: Maybe Text

The reason for the most recent state transition. This might be an empty string.

riitKeyName :: !Text

The key pair name, if this instance was launched with an associated key pair.

riitAmiLaunchIndex :: !Text

The AMI launch index, which can be used to find this instance in the launch group.

riitProductCodes :: [ProductCodesSetItemType]

The product codes attached to this instance.

riitInstanceType :: InstanceType

The instance type.

riitLaunchTime :: !UTCTime

The time the instance was launched.

riitPlacement :: !PlacementType

The location where the instance launched.

riitKernelId :: Maybe Text

The kernel associated with this instance.

riitRamdiskId :: Maybe Text

The RAM disk associated with this instance.

riitPlatform :: Maybe Text

The value is Windows for Windows AMIs; otherwise blank.

riitMonitoring :: !InstanceMonitoringStateType

The monitoring information for the instance.

riitSubnetId :: Maybe Text

The ID of the subnet in which the instance is running.

riitVpcId :: Maybe Text

The ID of the VPC in which the instance is running.

riitPrivateIpAddress :: Maybe Text

The private IP address assigned to the instance.

riitIpAddress :: Maybe Text

The IP address of the instance.

riitSourceDestCheck :: Maybe Bool

Specifies whether to enable an instance launched in a VPC to perform NAT. This controls whether source/destination checking is enabled on the instance. A value of true means checking is enabled, and false means checking is disabled. The value must be false for the instance to perform NAT. For more information, go to NAT Instances in the Amazon Virtual Private Cloud User Guide.

riitGroupSet :: [GroupItemType]

A list of the security groups for the instance.

riitStateReason :: Maybe StateReasonType

The reason for the most recent state transition. See StateReasonType for a listing of supported state change codes.

riitArchitecture :: !Text

The architecture of the image.

riitRootDeviceType :: !Text

The root device type used by the AMI. The AMI can use an Amazon EBS or instance store root device.

riitRootDeviceName :: Maybe Text

The root device name (for example, devsda1).

riitBlockDeviceMapping :: [InstanceBlockDeviceMappingResponseItemType]

Any block device mapping entries for the instance.

riitInstanceLifecycle :: Maybe Text

Indicates whether this is a Spot Instance.

riitSpotInstanceRequestId :: Maybe Text

The ID of the Spot Instance request.

riitVirtualizationType :: !Text

The instance's virtualization type.

riitClientToken :: Maybe Text

The idempotency token you provided when you launched the instance.

riitTagSet :: [ResourceTagSetItemType]

Any tags assigned to the resource.

riitHypervisor :: !Text

The instance's hypervisor type.

riitNetworkInterfaceSet :: [InstanceNetworkInterfaceSetItemType]

The network interfaces for the instance.

riitIamInstanceProfile :: Maybe IamInstanceProfileResponseType

The IAM Instance Profile (IIP) associated with the instance.

riitEbsOptimized :: !Bool

Indicates whether the instance is optimized for EBS I/O.

data SecurityGroupItemType

Constructors

SecurityGroupItemType

Fields

sgitOwnerId :: !Text

The AWS account ID of the owner of the security group.

sgitGroupId :: !Text

The ID of the security group.

sgitGroupName :: !Text

The name of the security group.

sgitGroupDescription :: !Text

A description of the security group.

sgitVpcId :: Maybe Text

EC2-VPC

The ID of the VPC for the security group.

sgitIpPermissions :: [IpPermissionType]

A list of inbound rules associated with the security group.

sgitIpPermissionsEgress :: [IpPermissionType]

EC2-VPC

A list of outbound rules associated with the security group. , sgitTagSet :: [ResourceTagSetItemType] -- ^ Any tags assigned to the resource, each one wrapped in an item element.

data StateReasonType

Constructors

StateReasonType

Fields

srtCode :: !Text

The reason code for the state change.

srtMessage :: !Text

The message for the state change.

data TagSetItemType

Constructors

TagSetItemType

Fields

tsitResourceId :: !Text

The ID of the resource. For example, ami-1a2b3c4d.

tsitResourceType :: !Text

The type of resource.

tsitKey :: !Text

The key of the tag.

tsitValue :: !Text

The value of the tag.

data Filter

Constructors

Filter

Fields

filterName :: !Text

filterValue :: [Text]

newtype Filters a

Constructors

Filters

Fields

unFilters :: [a]

data TagResourceType

Constructors

CustomerGateway

DhcpOptions

Image

Instance

InternetGateway

NetworkAcl

NetworkInterface

ReservedInstances

RouteTable

SecurityGroup

Snapshot

SpotInstancesRequest

Subnet

Volume

Vpc

VpnConnection

VpnGateway

data TagFilter

Constructors

TagKey [Text]
The tag key.

TagResourceId [Text]
The resource ID.

TagResourceType [TagResourceType]
The resource type.

TagValue [Text]
The tag value.

ec2NS :: ByteString

XML namespace to annotate EC2 elements with.

ec2Elem :: ByteString -> NName ByteString

Helper to define EC2 namespaced XML elements.

ec2XML :: XMLGeneric a

ec2ItemXML :: XMLGeneric a

Common

module Network.AWS