Network.AWS.EC2.AcceptVpcEndpointConnections

Description

Accepts one or more interface VPC endpoint connection requests to your VPC endpoint service.

Creating a Request

data AcceptVpcEndpointConnections

See: newAcceptVpcEndpointConnections smart constructor.

Constructors

AcceptVpcEndpointConnections'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

serviceId :: Text

The ID of the VPC endpoint service.

vpcEndpointIds :: [Text]

The IDs of one or more interface VPC endpoints.

newAcceptVpcEndpointConnections

Arguments

:: Text
$sel:serviceId:AcceptVpcEndpointConnections'

-> AcceptVpcEndpointConnections

Create a value of AcceptVpcEndpointConnections with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:AcceptVpcEndpointConnections', acceptVpcEndpointConnections_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:serviceId:AcceptVpcEndpointConnections', acceptVpcEndpointConnections_serviceId - The ID of the VPC endpoint service.

$sel:vpcEndpointIds:AcceptVpcEndpointConnections', acceptVpcEndpointConnections_vpcEndpointIds - The IDs of one or more interface VPC endpoints.

Request Lenses

acceptVpcEndpointConnections_dryRun :: Lens' AcceptVpcEndpointConnections (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

acceptVpcEndpointConnections_serviceId :: Lens' AcceptVpcEndpointConnections Text

The ID of the VPC endpoint service.

acceptVpcEndpointConnections_vpcEndpointIds :: Lens' AcceptVpcEndpointConnections [Text]

The IDs of one or more interface VPC endpoints.

Destructuring the Response

data AcceptVpcEndpointConnectionsResponse

See: newAcceptVpcEndpointConnectionsResponse smart constructor.

Constructors

AcceptVpcEndpointConnectionsResponse'

Fields

unsuccessful :: Maybe [UnsuccessfulItem]

Information about the interface endpoints that were not accepted, if applicable.

httpStatus :: Int

The response's http status code.

newAcceptVpcEndpointConnectionsResponse

Arguments

:: Int
$sel:httpStatus:AcceptVpcEndpointConnectionsResponse'

-> AcceptVpcEndpointConnectionsResponse

Create a value of AcceptVpcEndpointConnectionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:unsuccessful:AcceptVpcEndpointConnectionsResponse', acceptVpcEndpointConnectionsResponse_unsuccessful - Information about the interface endpoints that were not accepted, if applicable.

$sel:httpStatus:AcceptVpcEndpointConnectionsResponse', acceptVpcEndpointConnectionsResponse_httpStatus - The response's http status code.

Response Lenses

acceptVpcEndpointConnectionsResponse_unsuccessful :: Lens' AcceptVpcEndpointConnectionsResponse (Maybe [UnsuccessfulItem])

Information about the interface endpoints that were not accepted, if applicable.

acceptVpcEndpointConnectionsResponse_httpStatus :: Lens' AcceptVpcEndpointConnectionsResponse Int

The response's http status code.