Network.AWS.EC2.AssignIpv6Addresses

Description

Assigns one or more IPv6 addresses to the specified network interface. You can specify one or more specific IPv6 addresses, or you can specify the number of IPv6 addresses to be automatically assigned from within the subnet's IPv6 CIDR block range. You can assign as many IPv6 addresses to a network interface as you can assign private IPv4 addresses, and the limit varies per instance type. For information, see IP Addresses Per Network Interface Per Instance Type in the Amazon Elastic Compute Cloud User Guide.

You must specify either the IPv6 addresses or the IPv6 address count in the request.

You can optionally use Prefix Delegation on the network interface. You must specify either the IPV6 Prefix Delegation prefixes, or the IPv6 Prefix Delegation count. For information, see Assigning prefixes to Amazon EC2 network interfaces in the Amazon Elastic Compute Cloud User Guide.

Creating a Request

data AssignIpv6Addresses

See: newAssignIpv6Addresses smart constructor.

Constructors

AssignIpv6Addresses'

Fields

ipv6AddressCount :: Maybe Int

The number of additional IPv6 addresses to assign to the network interface. The specified number of IPv6 addresses are assigned in addition to the existing IPv6 addresses that are already assigned to the network interface. Amazon EC2 automatically selects the IPv6 addresses from the subnet range. You can't use this option if specifying specific IPv6 addresses.

ipv6Prefixes :: Maybe [Text]

One or more IPv6 prefixes assigned to the network interface. You cannot use this option if you use the Ipv6PrefixCount option.

ipv6PrefixCount :: Maybe Int

The number of IPv6 prefixes that Amazon Web Services automatically assigns to the network interface. You cannot use this option if you use the Ipv6Prefixes option.

ipv6Addresses :: Maybe [Text]

One or more specific IPv6 addresses to be assigned to the network interface. You can't use this option if you're specifying a number of IPv6 addresses.

networkInterfaceId :: Text

The ID of the network interface.

newAssignIpv6Addresses

Arguments

:: Text
$sel:networkInterfaceId:AssignIpv6Addresses'

-> AssignIpv6Addresses

Create a value of AssignIpv6Addresses with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ipv6AddressCount:AssignIpv6Addresses', assignIpv6Addresses_ipv6AddressCount - The number of additional IPv6 addresses to assign to the network interface. The specified number of IPv6 addresses are assigned in addition to the existing IPv6 addresses that are already assigned to the network interface. Amazon EC2 automatically selects the IPv6 addresses from the subnet range. You can't use this option if specifying specific IPv6 addresses.

$sel:ipv6Prefixes:AssignIpv6Addresses', assignIpv6Addresses_ipv6Prefixes - One or more IPv6 prefixes assigned to the network interface. You cannot use this option if you use the Ipv6PrefixCount option.

$sel:ipv6PrefixCount:AssignIpv6Addresses', assignIpv6Addresses_ipv6PrefixCount - The number of IPv6 prefixes that Amazon Web Services automatically assigns to the network interface. You cannot use this option if you use the Ipv6Prefixes option.

$sel:ipv6Addresses:AssignIpv6Addresses', assignIpv6Addresses_ipv6Addresses - One or more specific IPv6 addresses to be assigned to the network interface. You can't use this option if you're specifying a number of IPv6 addresses.

$sel:networkInterfaceId:AssignIpv6Addresses', assignIpv6Addresses_networkInterfaceId - The ID of the network interface.

Request Lenses

assignIpv6Addresses_ipv6AddressCount :: Lens' AssignIpv6Addresses (Maybe Int)

The number of additional IPv6 addresses to assign to the network interface. The specified number of IPv6 addresses are assigned in addition to the existing IPv6 addresses that are already assigned to the network interface. Amazon EC2 automatically selects the IPv6 addresses from the subnet range. You can't use this option if specifying specific IPv6 addresses.

assignIpv6Addresses_ipv6Prefixes :: Lens' AssignIpv6Addresses (Maybe [Text])

One or more IPv6 prefixes assigned to the network interface. You cannot use this option if you use the Ipv6PrefixCount option.

assignIpv6Addresses_ipv6PrefixCount :: Lens' AssignIpv6Addresses (Maybe Int)

The number of IPv6 prefixes that Amazon Web Services automatically assigns to the network interface. You cannot use this option if you use the Ipv6Prefixes option.

assignIpv6Addresses_ipv6Addresses :: Lens' AssignIpv6Addresses (Maybe [Text])

One or more specific IPv6 addresses to be assigned to the network interface. You can't use this option if you're specifying a number of IPv6 addresses.

assignIpv6Addresses_networkInterfaceId :: Lens' AssignIpv6Addresses Text

The ID of the network interface.

Destructuring the Response

data AssignIpv6AddressesResponse

See: newAssignIpv6AddressesResponse smart constructor.

Constructors

AssignIpv6AddressesResponse'

Fields

networkInterfaceId :: Maybe Text

The ID of the network interface.

assignedIpv6Prefixes :: Maybe [Text]

The IPv6 prefixes that are assigned to the network interface.

assignedIpv6Addresses :: Maybe [Text]

The new IPv6 addresses assigned to the network interface. Existing IPv6 addresses that were assigned to the network interface before the request are not included.

httpStatus :: Int

The response's http status code.

newAssignIpv6AddressesResponse

Arguments

:: Int
$sel:httpStatus:AssignIpv6AddressesResponse'

-> AssignIpv6AddressesResponse

Create a value of AssignIpv6AddressesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:networkInterfaceId:AssignIpv6Addresses', assignIpv6AddressesResponse_networkInterfaceId - The ID of the network interface.

$sel:assignedIpv6Prefixes:AssignIpv6AddressesResponse', assignIpv6AddressesResponse_assignedIpv6Prefixes - The IPv6 prefixes that are assigned to the network interface.

$sel:assignedIpv6Addresses:AssignIpv6AddressesResponse', assignIpv6AddressesResponse_assignedIpv6Addresses - The new IPv6 addresses assigned to the network interface. Existing IPv6 addresses that were assigned to the network interface before the request are not included.

$sel:httpStatus:AssignIpv6AddressesResponse', assignIpv6AddressesResponse_httpStatus - The response's http status code.

Response Lenses

assignIpv6AddressesResponse_networkInterfaceId :: Lens' AssignIpv6AddressesResponse (Maybe Text)

The ID of the network interface.

assignIpv6AddressesResponse_assignedIpv6Prefixes :: Lens' AssignIpv6AddressesResponse (Maybe [Text])

The IPv6 prefixes that are assigned to the network interface.

assignIpv6AddressesResponse_assignedIpv6Addresses :: Lens' AssignIpv6AddressesResponse (Maybe [Text])

The new IPv6 addresses assigned to the network interface. Existing IPv6 addresses that were assigned to the network interface before the request are not included.

assignIpv6AddressesResponse_httpStatus :: Lens' AssignIpv6AddressesResponse Int

The response's http status code.