Network.AWS.EC2.AssignPrivateIpAddresses

Description

Assigns one or more secondary private IP addresses to the specified network interface.

You can specify one or more specific secondary IP addresses, or you can specify the number of secondary IP addresses to be automatically assigned within the subnet's CIDR block range. The number of secondary IP addresses that you can assign to an instance varies by instance type. For information about instance types, see Instance Types in the Amazon Elastic Compute Cloud User Guide. For more information about Elastic IP addresses, see Elastic IP Addresses in the Amazon Elastic Compute Cloud User Guide.

When you move a secondary private IP address to another network interface, any Elastic IP address that is associated with the IP address is also moved.

Remapping an IP address is an asynchronous operation. When you move an IP address from one network interface to another, check network/interfaces/macs/mac/local-ipv4s in the instance metadata to confirm that the remapping is complete.

You must specify either the IP addresses or the IP address count in the request.

You can optionally use Prefix Delegation on the network interface. You must specify either the IPv4 Prefix Delegation prefixes, or the IPv4 Prefix Delegation count. For information, see Assigning prefixes to Amazon EC2 network interfaces in the Amazon Elastic Compute Cloud User Guide.

Creating a Request

data AssignPrivateIpAddresses

Contains the parameters for AssignPrivateIpAddresses.

See: newAssignPrivateIpAddresses smart constructor.

Constructors

AssignPrivateIpAddresses'

Fields

privateIpAddresses :: Maybe [Text]

One or more IP addresses to be assigned as a secondary private IP address to the network interface. You can't specify this parameter when also specifying a number of secondary IP addresses.

If you don't specify an IP address, Amazon EC2 automatically selects an IP address within the subnet range.

secondaryPrivateIpAddressCount :: Maybe Int

The number of secondary IP addresses to assign to the network interface. You can't specify this parameter when also specifying private IP addresses.

ipv4PrefixCount :: Maybe Int

The number of IPv4 prefixes that Amazon Web Services automatically assigns to the network interface. You cannot use this option if you use the Ipv4 Prefixes option.

allowReassignment :: Maybe Bool

Indicates whether to allow an IP address that is already assigned to another network interface or instance to be reassigned to the specified network interface.

ipv4Prefixes :: Maybe [Text]

One or more IPv4 prefixes assigned to the network interface. You cannot use this option if you use the Ipv4PrefixCount option.

networkInterfaceId :: Text

The ID of the network interface.

newAssignPrivateIpAddresses

Arguments

:: Text
$sel:networkInterfaceId:AssignPrivateIpAddresses'

-> AssignPrivateIpAddresses

Create a value of AssignPrivateIpAddresses with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:privateIpAddresses:AssignPrivateIpAddresses', assignPrivateIpAddresses_privateIpAddresses - One or more IP addresses to be assigned as a secondary private IP address to the network interface. You can't specify this parameter when also specifying a number of secondary IP addresses.

If you don't specify an IP address, Amazon EC2 automatically selects an IP address within the subnet range.

$sel:secondaryPrivateIpAddressCount:AssignPrivateIpAddresses', assignPrivateIpAddresses_secondaryPrivateIpAddressCount - The number of secondary IP addresses to assign to the network interface. You can't specify this parameter when also specifying private IP addresses.

$sel:ipv4PrefixCount:AssignPrivateIpAddresses', assignPrivateIpAddresses_ipv4PrefixCount - The number of IPv4 prefixes that Amazon Web Services automatically assigns to the network interface. You cannot use this option if you use the Ipv4 Prefixes option.

$sel:allowReassignment:AssignPrivateIpAddresses', assignPrivateIpAddresses_allowReassignment - Indicates whether to allow an IP address that is already assigned to another network interface or instance to be reassigned to the specified network interface.

$sel:ipv4Prefixes:AssignPrivateIpAddresses', assignPrivateIpAddresses_ipv4Prefixes - One or more IPv4 prefixes assigned to the network interface. You cannot use this option if you use the Ipv4PrefixCount option.

$sel:networkInterfaceId:AssignPrivateIpAddresses', assignPrivateIpAddresses_networkInterfaceId - The ID of the network interface.

Request Lenses

assignPrivateIpAddresses_privateIpAddresses :: Lens' AssignPrivateIpAddresses (Maybe [Text])

One or more IP addresses to be assigned as a secondary private IP address to the network interface. You can't specify this parameter when also specifying a number of secondary IP addresses.

If you don't specify an IP address, Amazon EC2 automatically selects an IP address within the subnet range.

assignPrivateIpAddresses_secondaryPrivateIpAddressCount :: Lens' AssignPrivateIpAddresses (Maybe Int)

The number of secondary IP addresses to assign to the network interface. You can't specify this parameter when also specifying private IP addresses.

assignPrivateIpAddresses_ipv4PrefixCount :: Lens' AssignPrivateIpAddresses (Maybe Int)

The number of IPv4 prefixes that Amazon Web Services automatically assigns to the network interface. You cannot use this option if you use the Ipv4 Prefixes option.

assignPrivateIpAddresses_allowReassignment :: Lens' AssignPrivateIpAddresses (Maybe Bool)

Indicates whether to allow an IP address that is already assigned to another network interface or instance to be reassigned to the specified network interface.

assignPrivateIpAddresses_ipv4Prefixes :: Lens' AssignPrivateIpAddresses (Maybe [Text])

One or more IPv4 prefixes assigned to the network interface. You cannot use this option if you use the Ipv4PrefixCount option.

assignPrivateIpAddresses_networkInterfaceId :: Lens' AssignPrivateIpAddresses Text

The ID of the network interface.

Destructuring the Response

data AssignPrivateIpAddressesResponse

See: newAssignPrivateIpAddressesResponse smart constructor.

Constructors

AssignPrivateIpAddressesResponse'

Fields

assignedPrivateIpAddresses :: Maybe [AssignedPrivateIpAddress]

The private IP addresses assigned to the network interface.

networkInterfaceId :: Maybe Text

The ID of the network interface.

assignedIpv4Prefixes :: Maybe [Ipv4PrefixSpecification]

The IPv4 prefixes that are assigned to the network interface.

httpStatus :: Int

The response's http status code.

newAssignPrivateIpAddressesResponse

Arguments

:: Int
$sel:httpStatus:AssignPrivateIpAddressesResponse'

-> AssignPrivateIpAddressesResponse

Create a value of AssignPrivateIpAddressesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:assignedPrivateIpAddresses:AssignPrivateIpAddressesResponse', assignPrivateIpAddressesResponse_assignedPrivateIpAddresses - The private IP addresses assigned to the network interface.

$sel:networkInterfaceId:AssignPrivateIpAddresses', assignPrivateIpAddressesResponse_networkInterfaceId - The ID of the network interface.

$sel:assignedIpv4Prefixes:AssignPrivateIpAddressesResponse', assignPrivateIpAddressesResponse_assignedIpv4Prefixes - The IPv4 prefixes that are assigned to the network interface.

$sel:httpStatus:AssignPrivateIpAddressesResponse', assignPrivateIpAddressesResponse_httpStatus - The response's http status code.

Response Lenses

assignPrivateIpAddressesResponse_assignedPrivateIpAddresses :: Lens' AssignPrivateIpAddressesResponse (Maybe [AssignedPrivateIpAddress])

The private IP addresses assigned to the network interface.

assignPrivateIpAddressesResponse_networkInterfaceId :: Lens' AssignPrivateIpAddressesResponse (Maybe Text)

The ID of the network interface.

assignPrivateIpAddressesResponse_assignedIpv4Prefixes :: Lens' AssignPrivateIpAddressesResponse (Maybe [Ipv4PrefixSpecification])

The IPv4 prefixes that are assigned to the network interface.

assignPrivateIpAddressesResponse_httpStatus :: Lens' AssignPrivateIpAddressesResponse Int

The response's http status code.