Network.AWS.EC2.AssociateClientVpnTargetNetwork

Description

Associates a target network with a Client VPN endpoint. A target network is a subnet in a VPC. You can associate multiple subnets from the same VPC with a Client VPN endpoint. You can associate only one subnet in each Availability Zone. We recommend that you associate at least two subnets to provide Availability Zone redundancy.

If you specified a VPC when you created the Client VPN endpoint or if you have previous subnet associations, the specified subnet must be in the same VPC. To specify a subnet that's in a different VPC, you must first modify the Client VPN endpoint (ModifyClientVpnEndpoint) and change the VPC that's associated with it.

Creating a Request

data AssociateClientVpnTargetNetwork

See: newAssociateClientVpnTargetNetwork smart constructor.

Constructors

AssociateClientVpnTargetNetwork'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

clientToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to ensure idempotency.

clientVpnEndpointId :: Text

The ID of the Client VPN endpoint.

subnetId :: Text

The ID of the subnet to associate with the Client VPN endpoint.

newAssociateClientVpnTargetNetwork

Arguments

:: Text
$sel:clientVpnEndpointId:AssociateClientVpnTargetNetwork'

-> Text
$sel:subnetId:AssociateClientVpnTargetNetwork'

-> AssociateClientVpnTargetNetwork

Create a value of AssociateClientVpnTargetNetwork with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:AssociateClientVpnTargetNetwork', associateClientVpnTargetNetwork_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:clientToken:AssociateClientVpnTargetNetwork', associateClientVpnTargetNetwork_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to ensure idempotency.

$sel:clientVpnEndpointId:AssociateClientVpnTargetNetwork', associateClientVpnTargetNetwork_clientVpnEndpointId - The ID of the Client VPN endpoint.

$sel:subnetId:AssociateClientVpnTargetNetwork', associateClientVpnTargetNetwork_subnetId - The ID of the subnet to associate with the Client VPN endpoint.

Request Lenses

associateClientVpnTargetNetwork_dryRun :: Lens' AssociateClientVpnTargetNetwork (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

associateClientVpnTargetNetwork_clientToken :: Lens' AssociateClientVpnTargetNetwork (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to ensure idempotency.

associateClientVpnTargetNetwork_clientVpnEndpointId :: Lens' AssociateClientVpnTargetNetwork Text

The ID of the Client VPN endpoint.

associateClientVpnTargetNetwork_subnetId :: Lens' AssociateClientVpnTargetNetwork Text

The ID of the subnet to associate with the Client VPN endpoint.

Destructuring the Response

data AssociateClientVpnTargetNetworkResponse

See: newAssociateClientVpnTargetNetworkResponse smart constructor.

Constructors

AssociateClientVpnTargetNetworkResponse'

Fields

status :: Maybe AssociationStatus

The current state of the target network association.

associationId :: Maybe Text

The unique ID of the target network association.

httpStatus :: Int

The response's http status code.

newAssociateClientVpnTargetNetworkResponse

Arguments

:: Int
$sel:httpStatus:AssociateClientVpnTargetNetworkResponse'

-> AssociateClientVpnTargetNetworkResponse

Create a value of AssociateClientVpnTargetNetworkResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:AssociateClientVpnTargetNetworkResponse', associateClientVpnTargetNetworkResponse_status - The current state of the target network association.

$sel:associationId:AssociateClientVpnTargetNetworkResponse', associateClientVpnTargetNetworkResponse_associationId - The unique ID of the target network association.

$sel:httpStatus:AssociateClientVpnTargetNetworkResponse', associateClientVpnTargetNetworkResponse_httpStatus - The response's http status code.

Response Lenses

associateClientVpnTargetNetworkResponse_status :: Lens' AssociateClientVpnTargetNetworkResponse (Maybe AssociationStatus)

The current state of the target network association.

associateClientVpnTargetNetworkResponse_associationId :: Lens' AssociateClientVpnTargetNetworkResponse (Maybe Text)

The unique ID of the target network association.

associateClientVpnTargetNetworkResponse_httpStatus :: Lens' AssociateClientVpnTargetNetworkResponse Int

The response's http status code.