Network.AWS.EC2.AssociateEnclaveCertificateIamRole

Description

Associates an Identity and Access Management (IAM) role with an Certificate Manager (ACM) certificate. This enables the certificate to be used by the ACM for Nitro Enclaves application inside an enclave. For more information, see Certificate Manager for Nitro Enclaves in the Amazon Web Services Nitro Enclaves User Guide.

When the IAM role is associated with the ACM certificate, the certificate, certificate chain, and encrypted private key are placed in an Amazon S3 bucket that only the associated IAM role can access. The private key of the certificate is encrypted with an Amazon Web Services managed key that has an attached attestation-based key policy.

To enable the IAM role to access the Amazon S3 object, you must grant it permission to call s3:GetObject on the Amazon S3 bucket returned by the command. To enable the IAM role to access the KMS key, you must grant it permission to call kms:Decrypt on the KMS key returned by the command. For more information, see Grant the role permission to access the certificate and encryption key in the Amazon Web Services Nitro Enclaves User Guide.

Creating a Request

data AssociateEnclaveCertificateIamRole

See: newAssociateEnclaveCertificateIamRole smart constructor.

Constructors

AssociateEnclaveCertificateIamRole'

Fields

roleArn :: Maybe Text

The ARN of the IAM role to associate with the ACM certificate. You can associate up to 16 IAM roles with an ACM certificate.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

certificateArn :: Maybe Text

The ARN of the ACM certificate with which to associate the IAM role.

newAssociateEnclaveCertificateIamRole :: AssociateEnclaveCertificateIamRole

Create a value of AssociateEnclaveCertificateIamRole with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:roleArn:AssociateEnclaveCertificateIamRole', associateEnclaveCertificateIamRole_roleArn - The ARN of the IAM role to associate with the ACM certificate. You can associate up to 16 IAM roles with an ACM certificate.

$sel:dryRun:AssociateEnclaveCertificateIamRole', associateEnclaveCertificateIamRole_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:certificateArn:AssociateEnclaveCertificateIamRole', associateEnclaveCertificateIamRole_certificateArn - The ARN of the ACM certificate with which to associate the IAM role.

Request Lenses

associateEnclaveCertificateIamRole_roleArn :: Lens' AssociateEnclaveCertificateIamRole (Maybe Text)

The ARN of the IAM role to associate with the ACM certificate. You can associate up to 16 IAM roles with an ACM certificate.

associateEnclaveCertificateIamRole_dryRun :: Lens' AssociateEnclaveCertificateIamRole (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

associateEnclaveCertificateIamRole_certificateArn :: Lens' AssociateEnclaveCertificateIamRole (Maybe Text)

The ARN of the ACM certificate with which to associate the IAM role.

Destructuring the Response

data AssociateEnclaveCertificateIamRoleResponse

See: newAssociateEnclaveCertificateIamRoleResponse smart constructor.

Constructors

AssociateEnclaveCertificateIamRoleResponse'

Fields

certificateS3ObjectKey :: Maybe Text

The Amazon S3 object key where the certificate, certificate chain, and encrypted private key bundle are stored. The object key is formatted as follows: role_arn/certificate_arn.

encryptionKmsKeyId :: Maybe Text

The ID of the KMS key used to encrypt the private key of the certificate.

certificateS3BucketName :: Maybe Text

The name of the Amazon S3 bucket to which the certificate was uploaded.

httpStatus :: Int

The response's http status code.

newAssociateEnclaveCertificateIamRoleResponse

Arguments

:: Int
$sel:httpStatus:AssociateEnclaveCertificateIamRoleResponse'

-> AssociateEnclaveCertificateIamRoleResponse

Create a value of AssociateEnclaveCertificateIamRoleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:certificateS3ObjectKey:AssociateEnclaveCertificateIamRoleResponse', associateEnclaveCertificateIamRoleResponse_certificateS3ObjectKey - The Amazon S3 object key where the certificate, certificate chain, and encrypted private key bundle are stored. The object key is formatted as follows: role_arn/certificate_arn.

$sel:encryptionKmsKeyId:AssociateEnclaveCertificateIamRoleResponse', associateEnclaveCertificateIamRoleResponse_encryptionKmsKeyId - The ID of the KMS key used to encrypt the private key of the certificate.

$sel:certificateS3BucketName:AssociateEnclaveCertificateIamRoleResponse', associateEnclaveCertificateIamRoleResponse_certificateS3BucketName - The name of the Amazon S3 bucket to which the certificate was uploaded.

$sel:httpStatus:AssociateEnclaveCertificateIamRoleResponse', associateEnclaveCertificateIamRoleResponse_httpStatus - The response's http status code.

Response Lenses

associateEnclaveCertificateIamRoleResponse_certificateS3ObjectKey :: Lens' AssociateEnclaveCertificateIamRoleResponse (Maybe Text)

The Amazon S3 object key where the certificate, certificate chain, and encrypted private key bundle are stored. The object key is formatted as follows: role_arn/certificate_arn.

associateEnclaveCertificateIamRoleResponse_encryptionKmsKeyId :: Lens' AssociateEnclaveCertificateIamRoleResponse (Maybe Text)

The ID of the KMS key used to encrypt the private key of the certificate.

associateEnclaveCertificateIamRoleResponse_certificateS3BucketName :: Lens' AssociateEnclaveCertificateIamRoleResponse (Maybe Text)

The name of the Amazon S3 bucket to which the certificate was uploaded.

associateEnclaveCertificateIamRoleResponse_httpStatus :: Lens' AssociateEnclaveCertificateIamRoleResponse Int

The response's http status code.