Network.AWS.EC2.AssociateRouteTable

Description

Associates a subnet in your VPC or an internet gateway or virtual private gateway attached to your VPC with a route table in your VPC. This association causes traffic from the subnet or gateway to be routed according to the routes in the route table. The action returns an association ID, which you need in order to disassociate the route table later. A route table can be associated with multiple subnets.

For more information, see Route tables in the Amazon Virtual Private Cloud User Guide.

Creating a Request

data AssociateRouteTable

See: newAssociateRouteTable smart constructor.

Constructors

AssociateRouteTable'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

subnetId :: Maybe Text

The ID of the subnet.

gatewayId :: Maybe Text

The ID of the internet gateway or virtual private gateway.

routeTableId :: Text

The ID of the route table.

newAssociateRouteTable

Arguments

:: Text
$sel:routeTableId:AssociateRouteTable'

-> AssociateRouteTable

Create a value of AssociateRouteTable with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:AssociateRouteTable', associateRouteTable_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:subnetId:AssociateRouteTable', associateRouteTable_subnetId - The ID of the subnet.

$sel:gatewayId:AssociateRouteTable', associateRouteTable_gatewayId - The ID of the internet gateway or virtual private gateway.

$sel:routeTableId:AssociateRouteTable', associateRouteTable_routeTableId - The ID of the route table.

Request Lenses

associateRouteTable_dryRun :: Lens' AssociateRouteTable (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

associateRouteTable_subnetId :: Lens' AssociateRouteTable (Maybe Text)

The ID of the subnet.

associateRouteTable_gatewayId :: Lens' AssociateRouteTable (Maybe Text)

The ID of the internet gateway or virtual private gateway.

associateRouteTable_routeTableId :: Lens' AssociateRouteTable Text

The ID of the route table.

Destructuring the Response

data AssociateRouteTableResponse

See: newAssociateRouteTableResponse smart constructor.

Constructors

AssociateRouteTableResponse'

Fields

associationState :: Maybe RouteTableAssociationState

The state of the association.

associationId :: Maybe Text

The route table association ID. This ID is required for disassociating the route table.

httpStatus :: Int

The response's http status code.

newAssociateRouteTableResponse

Arguments

:: Int
$sel:httpStatus:AssociateRouteTableResponse'

-> AssociateRouteTableResponse

Create a value of AssociateRouteTableResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:associationState:AssociateRouteTableResponse', associateRouteTableResponse_associationState - The state of the association.

$sel:associationId:AssociateRouteTableResponse', associateRouteTableResponse_associationId - The route table association ID. This ID is required for disassociating the route table.

$sel:httpStatus:AssociateRouteTableResponse', associateRouteTableResponse_httpStatus - The response's http status code.

Response Lenses

associateRouteTableResponse_associationState :: Lens' AssociateRouteTableResponse (Maybe RouteTableAssociationState)

The state of the association.

associateRouteTableResponse_associationId :: Lens' AssociateRouteTableResponse (Maybe Text)

The route table association ID. This ID is required for disassociating the route table.

associateRouteTableResponse_httpStatus :: Lens' AssociateRouteTableResponse Int

The response's http status code.