Network.AWS.EC2.AssociateTransitGatewayMulticastDomain

Description

Associates the specified subnets and transit gateway attachments with the specified transit gateway multicast domain.

The transit gateway attachment must be in the available state before you can add a resource. Use DescribeTransitGatewayAttachments to see the state of the attachment.

Creating a Request

data AssociateTransitGatewayMulticastDomain

See: newAssociateTransitGatewayMulticastDomain smart constructor.

Constructors

AssociateTransitGatewayMulticastDomain'

Fields

transitGatewayMulticastDomainId :: Maybe Text

The ID of the transit gateway multicast domain.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

subnetIds :: Maybe [Text]

The IDs of the subnets to associate with the transit gateway multicast domain.

transitGatewayAttachmentId :: Maybe Text

The ID of the transit gateway attachment to associate with the transit gateway multicast domain.

newAssociateTransitGatewayMulticastDomain :: AssociateTransitGatewayMulticastDomain

Create a value of AssociateTransitGatewayMulticastDomain with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:transitGatewayMulticastDomainId:AssociateTransitGatewayMulticastDomain', associateTransitGatewayMulticastDomain_transitGatewayMulticastDomainId - The ID of the transit gateway multicast domain.

$sel:dryRun:AssociateTransitGatewayMulticastDomain', associateTransitGatewayMulticastDomain_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:subnetIds:AssociateTransitGatewayMulticastDomain', associateTransitGatewayMulticastDomain_subnetIds - The IDs of the subnets to associate with the transit gateway multicast domain.

$sel:transitGatewayAttachmentId:AssociateTransitGatewayMulticastDomain', associateTransitGatewayMulticastDomain_transitGatewayAttachmentId - The ID of the transit gateway attachment to associate with the transit gateway multicast domain.

Request Lenses

associateTransitGatewayMulticastDomain_transitGatewayMulticastDomainId :: Lens' AssociateTransitGatewayMulticastDomain (Maybe Text)

The ID of the transit gateway multicast domain.

associateTransitGatewayMulticastDomain_dryRun :: Lens' AssociateTransitGatewayMulticastDomain (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

associateTransitGatewayMulticastDomain_subnetIds :: Lens' AssociateTransitGatewayMulticastDomain (Maybe [Text])

The IDs of the subnets to associate with the transit gateway multicast domain.

associateTransitGatewayMulticastDomain_transitGatewayAttachmentId :: Lens' AssociateTransitGatewayMulticastDomain (Maybe Text)

The ID of the transit gateway attachment to associate with the transit gateway multicast domain.

Destructuring the Response

data AssociateTransitGatewayMulticastDomainResponse

See: newAssociateTransitGatewayMulticastDomainResponse smart constructor.

Constructors

AssociateTransitGatewayMulticastDomainResponse'

Fields

associations :: Maybe TransitGatewayMulticastDomainAssociations

Information about the transit gateway multicast domain associations.

httpStatus :: Int

The response's http status code.

newAssociateTransitGatewayMulticastDomainResponse

Arguments

:: Int
$sel:httpStatus:AssociateTransitGatewayMulticastDomainResponse'

-> AssociateTransitGatewayMulticastDomainResponse

Create a value of AssociateTransitGatewayMulticastDomainResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:associations:AssociateTransitGatewayMulticastDomainResponse', associateTransitGatewayMulticastDomainResponse_associations - Information about the transit gateway multicast domain associations.

$sel:httpStatus:AssociateTransitGatewayMulticastDomainResponse', associateTransitGatewayMulticastDomainResponse_httpStatus - The response's http status code.

Response Lenses

associateTransitGatewayMulticastDomainResponse_associations :: Lens' AssociateTransitGatewayMulticastDomainResponse (Maybe TransitGatewayMulticastDomainAssociations)

Information about the transit gateway multicast domain associations.

associateTransitGatewayMulticastDomainResponse_httpStatus :: Lens' AssociateTransitGatewayMulticastDomainResponse Int

The response's http status code.