Network.AWS.EC2.AssociateTrunkInterface

Description

This API action is currently in limited preview only. If you are interested in using this feature, contact your account manager.

Associates a branch network interface with a trunk network interface.

Before you create the association, run the create-network-interface command and set --interface-type to trunk. You must also create a network interface for each branch network interface that you want to associate with the trunk network interface.

Creating a Request

data AssociateTrunkInterface

See: newAssociateTrunkInterface smart constructor.

Constructors

AssociateTrunkInterface'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

greKey :: Maybe Int

The application key. This applies to the GRE protocol.

vlanId :: Maybe Int

The ID of the VLAN. This applies to the VLAN protocol.

clientToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to Ensure Idempotency.

branchInterfaceId :: Text

The ID of the branch network interface.

trunkInterfaceId :: Text

The ID of the trunk network interface.

newAssociateTrunkInterface

Arguments

:: Text
$sel:branchInterfaceId:AssociateTrunkInterface'

-> Text
$sel:trunkInterfaceId:AssociateTrunkInterface'

-> AssociateTrunkInterface

Create a value of AssociateTrunkInterface with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:AssociateTrunkInterface', associateTrunkInterface_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:greKey:AssociateTrunkInterface', associateTrunkInterface_greKey - The application key. This applies to the GRE protocol.

$sel:vlanId:AssociateTrunkInterface', associateTrunkInterface_vlanId - The ID of the VLAN. This applies to the VLAN protocol.

$sel:clientToken:AssociateTrunkInterface', associateTrunkInterface_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to Ensure Idempotency.

$sel:branchInterfaceId:AssociateTrunkInterface', associateTrunkInterface_branchInterfaceId - The ID of the branch network interface.

$sel:trunkInterfaceId:AssociateTrunkInterface', associateTrunkInterface_trunkInterfaceId - The ID of the trunk network interface.

Request Lenses

associateTrunkInterface_dryRun :: Lens' AssociateTrunkInterface (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

associateTrunkInterface_greKey :: Lens' AssociateTrunkInterface (Maybe Int)

The application key. This applies to the GRE protocol.

associateTrunkInterface_vlanId :: Lens' AssociateTrunkInterface (Maybe Int)

The ID of the VLAN. This applies to the VLAN protocol.

associateTrunkInterface_clientToken :: Lens' AssociateTrunkInterface (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to Ensure Idempotency.

associateTrunkInterface_branchInterfaceId :: Lens' AssociateTrunkInterface Text

The ID of the branch network interface.

associateTrunkInterface_trunkInterfaceId :: Lens' AssociateTrunkInterface Text

The ID of the trunk network interface.

Destructuring the Response

data AssociateTrunkInterfaceResponse

See: newAssociateTrunkInterfaceResponse smart constructor.

Constructors

AssociateTrunkInterfaceResponse'

Fields

interfaceAssociation :: Maybe TrunkInterfaceAssociation

Information about the association between the trunk network interface and branch network interface.

clientToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to Ensure Idempotency.

httpStatus :: Int

The response's http status code.

newAssociateTrunkInterfaceResponse

Arguments

:: Int
$sel:httpStatus:AssociateTrunkInterfaceResponse'

-> AssociateTrunkInterfaceResponse

Create a value of AssociateTrunkInterfaceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:interfaceAssociation:AssociateTrunkInterfaceResponse', associateTrunkInterfaceResponse_interfaceAssociation - Information about the association between the trunk network interface and branch network interface.

$sel:clientToken:AssociateTrunkInterface', associateTrunkInterfaceResponse_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to Ensure Idempotency.

$sel:httpStatus:AssociateTrunkInterfaceResponse', associateTrunkInterfaceResponse_httpStatus - The response's http status code.

Response Lenses

associateTrunkInterfaceResponse_interfaceAssociation :: Lens' AssociateTrunkInterfaceResponse (Maybe TrunkInterfaceAssociation)

Information about the association between the trunk network interface and branch network interface.

associateTrunkInterfaceResponse_clientToken :: Lens' AssociateTrunkInterfaceResponse (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to Ensure Idempotency.

associateTrunkInterfaceResponse_httpStatus :: Lens' AssociateTrunkInterfaceResponse Int

The response's http status code.