Network.AWS.EC2.AssociateVpcCidrBlock

Description

Associates a CIDR block with your VPC. You can associate a secondary IPv4 CIDR block, an Amazon-provided IPv6 CIDR block, or an IPv6 CIDR block from an IPv6 address pool that you provisioned through bring your own IP addresses (BYOIP). The IPv6 CIDR block size is fixed at /56.

You must specify one of the following in the request: an IPv4 CIDR block, an IPv6 pool, or an Amazon-provided IPv6 CIDR block.

For more information about associating CIDR blocks with your VPC and applicable restrictions, see VPC and subnet sizing in the Amazon Virtual Private Cloud User Guide.

Creating a Request

data AssociateVpcCidrBlock

See: newAssociateVpcCidrBlock smart constructor.

Constructors

AssociateVpcCidrBlock'

Fields

ipv6Pool :: Maybe Text

The ID of an IPv6 address pool from which to allocate the IPv6 CIDR block.

ipv6CidrBlock :: Maybe Text

An IPv6 CIDR block from the IPv6 address pool. You must also specify Ipv6Pool in the request.

To let Amazon choose the IPv6 CIDR block for you, omit this parameter.

ipv6CidrBlockNetworkBorderGroup :: Maybe Text

The name of the location from which we advertise the IPV6 CIDR block. Use this parameter to limit the CIDR block to this location.

You must set AmazonProvidedIpv6CidrBlock to true to use this parameter.

You can have one IPv6 CIDR block association per network border group.

amazonProvidedIpv6CidrBlock :: Maybe Bool

Requests an Amazon-provided IPv6 CIDR block with a /56 prefix length for the VPC. You cannot specify the range of IPv6 addresses, or the size of the CIDR block.

cidrBlock :: Maybe Text

An IPv4 CIDR block to associate with the VPC.

vpcId :: Text

The ID of the VPC.

newAssociateVpcCidrBlock

Arguments

:: Text
$sel:vpcId:AssociateVpcCidrBlock'

-> AssociateVpcCidrBlock

Create a value of AssociateVpcCidrBlock with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ipv6Pool:AssociateVpcCidrBlock', associateVpcCidrBlock_ipv6Pool - The ID of an IPv6 address pool from which to allocate the IPv6 CIDR block.

$sel:ipv6CidrBlock:AssociateVpcCidrBlock', associateVpcCidrBlock_ipv6CidrBlock - An IPv6 CIDR block from the IPv6 address pool. You must also specify Ipv6Pool in the request.

To let Amazon choose the IPv6 CIDR block for you, omit this parameter.

$sel:ipv6CidrBlockNetworkBorderGroup:AssociateVpcCidrBlock', associateVpcCidrBlock_ipv6CidrBlockNetworkBorderGroup - The name of the location from which we advertise the IPV6 CIDR block. Use this parameter to limit the CIDR block to this location.

You must set AmazonProvidedIpv6CidrBlock to true to use this parameter.

You can have one IPv6 CIDR block association per network border group.

$sel:amazonProvidedIpv6CidrBlock:AssociateVpcCidrBlock', associateVpcCidrBlock_amazonProvidedIpv6CidrBlock - Requests an Amazon-provided IPv6 CIDR block with a /56 prefix length for the VPC. You cannot specify the range of IPv6 addresses, or the size of the CIDR block.

$sel:cidrBlock:AssociateVpcCidrBlock', associateVpcCidrBlock_cidrBlock - An IPv4 CIDR block to associate with the VPC.

$sel:vpcId:AssociateVpcCidrBlock', associateVpcCidrBlock_vpcId - The ID of the VPC.

Request Lenses

associateVpcCidrBlock_ipv6Pool :: Lens' AssociateVpcCidrBlock (Maybe Text)

The ID of an IPv6 address pool from which to allocate the IPv6 CIDR block.

associateVpcCidrBlock_ipv6CidrBlock :: Lens' AssociateVpcCidrBlock (Maybe Text)

An IPv6 CIDR block from the IPv6 address pool. You must also specify Ipv6Pool in the request.

To let Amazon choose the IPv6 CIDR block for you, omit this parameter.

associateVpcCidrBlock_ipv6CidrBlockNetworkBorderGroup :: Lens' AssociateVpcCidrBlock (Maybe Text)

The name of the location from which we advertise the IPV6 CIDR block. Use this parameter to limit the CIDR block to this location.

You must set AmazonProvidedIpv6CidrBlock to true to use this parameter.

You can have one IPv6 CIDR block association per network border group.

associateVpcCidrBlock_amazonProvidedIpv6CidrBlock :: Lens' AssociateVpcCidrBlock (Maybe Bool)

Requests an Amazon-provided IPv6 CIDR block with a /56 prefix length for the VPC. You cannot specify the range of IPv6 addresses, or the size of the CIDR block.

associateVpcCidrBlock_cidrBlock :: Lens' AssociateVpcCidrBlock (Maybe Text)

An IPv4 CIDR block to associate with the VPC.

associateVpcCidrBlock_vpcId :: Lens' AssociateVpcCidrBlock Text

The ID of the VPC.

Destructuring the Response

data AssociateVpcCidrBlockResponse

See: newAssociateVpcCidrBlockResponse smart constructor.

Constructors

AssociateVpcCidrBlockResponse'

Fields

ipv6CidrBlockAssociation :: Maybe VpcIpv6CidrBlockAssociation

Information about the IPv6 CIDR block association.

cidrBlockAssociation :: Maybe VpcCidrBlockAssociation

Information about the IPv4 CIDR block association.

vpcId :: Maybe Text

The ID of the VPC.

httpStatus :: Int

The response's http status code.

newAssociateVpcCidrBlockResponse

Arguments

:: Int
$sel:httpStatus:AssociateVpcCidrBlockResponse'

-> AssociateVpcCidrBlockResponse

Create a value of AssociateVpcCidrBlockResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ipv6CidrBlockAssociation:AssociateVpcCidrBlockResponse', associateVpcCidrBlockResponse_ipv6CidrBlockAssociation - Information about the IPv6 CIDR block association.

$sel:cidrBlockAssociation:AssociateVpcCidrBlockResponse', associateVpcCidrBlockResponse_cidrBlockAssociation - Information about the IPv4 CIDR block association.

$sel:vpcId:AssociateVpcCidrBlock', associateVpcCidrBlockResponse_vpcId - The ID of the VPC.

$sel:httpStatus:AssociateVpcCidrBlockResponse', associateVpcCidrBlockResponse_httpStatus - The response's http status code.

Response Lenses

associateVpcCidrBlockResponse_ipv6CidrBlockAssociation :: Lens' AssociateVpcCidrBlockResponse (Maybe VpcIpv6CidrBlockAssociation)

Information about the IPv6 CIDR block association.

associateVpcCidrBlockResponse_cidrBlockAssociation :: Lens' AssociateVpcCidrBlockResponse (Maybe VpcCidrBlockAssociation)

Information about the IPv4 CIDR block association.

associateVpcCidrBlockResponse_vpcId :: Lens' AssociateVpcCidrBlockResponse (Maybe Text)

The ID of the VPC.

associateVpcCidrBlockResponse_httpStatus :: Lens' AssociateVpcCidrBlockResponse Int

The response's http status code.