Network.AWS.EC2.AttachVolume

Description

Attaches an EBS volume to a running or stopped instance and exposes it to the instance with the specified device name.

Encrypted EBS volumes must be attached to instances that support Amazon EBS encryption. For more information, see Amazon EBS encryption in the Amazon Elastic Compute Cloud User Guide.

After you attach an EBS volume, you must make it available. For more information, see Make an EBS volume available for use.

If a volume has an Amazon Web Services Marketplace product code:

The volume can be attached only to a stopped instance.

Amazon Web Services Marketplace product codes are copied from the volume to the instance.

You must be subscribed to the product.

The instance type and operating system of the instance must support the product. For example, you can't detach a volume from a Windows instance and attach it to a Linux instance.

For more information, see Attach an Amazon EBS volume to an instance in the Amazon Elastic Compute Cloud User Guide.

Creating a Request

data AttachVolume

See: newAttachVolume smart constructor.

Constructors

AttachVolume'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

device :: Text

The device name (for example, /dev/sdh or xvdh).

instanceId :: Text

The ID of the instance.

volumeId :: Text

The ID of the EBS volume. The volume and instance must be within the same Availability Zone.

newAttachVolume

Arguments

:: Text
$sel:device:AttachVolume'

-> Text
$sel:instanceId:AttachVolume'

-> Text
$sel:volumeId:AttachVolume'

-> AttachVolume

Create a value of AttachVolume with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:AttachVolume', attachVolume_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:device:AttachVolume', attachVolume_device - The device name (for example, /dev/sdh or xvdh).

$sel:instanceId:AttachVolume', attachVolume_instanceId - The ID of the instance.

$sel:volumeId:AttachVolume', attachVolume_volumeId - The ID of the EBS volume. The volume and instance must be within the same Availability Zone.

Request Lenses

attachVolume_dryRun :: Lens' AttachVolume (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

attachVolume_device :: Lens' AttachVolume Text

The device name (for example, /dev/sdh or xvdh).

attachVolume_instanceId :: Lens' AttachVolume Text

The ID of the instance.

attachVolume_volumeId :: Lens' AttachVolume Text

The ID of the EBS volume. The volume and instance must be within the same Availability Zone.

Destructuring the Response

data VolumeAttachment

Describes volume attachment details.

See: newVolumeAttachment smart constructor.

Constructors

VolumeAttachment'

Fields

instanceId :: Maybe Text

The ID of the instance.

deleteOnTermination :: Maybe Bool

Indicates whether the EBS volume is deleted on instance termination.

state :: Maybe VolumeAttachmentState

The attachment state of the volume.

device :: Maybe Text

The device name.

volumeId :: Maybe Text

The ID of the volume.

attachTime :: Maybe ISO8601

The time stamp when the attachment initiated.

newVolumeAttachment :: VolumeAttachment

Create a value of VolumeAttachment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceId:VolumeAttachment', volumeAttachment_instanceId - The ID of the instance.

$sel:deleteOnTermination:VolumeAttachment', volumeAttachment_deleteOnTermination - Indicates whether the EBS volume is deleted on instance termination.

$sel:state:VolumeAttachment', volumeAttachment_state - The attachment state of the volume.

$sel:device:VolumeAttachment', volumeAttachment_device - The device name.

$sel:volumeId:VolumeAttachment', volumeAttachment_volumeId - The ID of the volume.

$sel:attachTime:VolumeAttachment', volumeAttachment_attachTime - The time stamp when the attachment initiated.

Response Lenses

volumeAttachment_instanceId :: Lens' VolumeAttachment (Maybe Text)

The ID of the instance.

volumeAttachment_deleteOnTermination :: Lens' VolumeAttachment (Maybe Bool)

Indicates whether the EBS volume is deleted on instance termination.

volumeAttachment_state :: Lens' VolumeAttachment (Maybe VolumeAttachmentState)

The attachment state of the volume.

volumeAttachment_device :: Lens' VolumeAttachment (Maybe Text)

The device name.

volumeAttachment_volumeId :: Lens' VolumeAttachment (Maybe Text)

The ID of the volume.

volumeAttachment_attachTime :: Lens' VolumeAttachment (Maybe UTCTime)

The time stamp when the attachment initiated.