Network.AWS.EC2.AuthorizeClientVpnIngress

Description

Adds an ingress authorization rule to a Client VPN endpoint. Ingress authorization rules act as firewall rules that grant access to networks. You must configure ingress authorization rules to enable clients to access resources in Amazon Web Services or on-premises networks.

Creating a Request

data AuthorizeClientVpnIngress

See: newAuthorizeClientVpnIngress smart constructor.

Constructors

AuthorizeClientVpnIngress'

Fields

accessGroupId :: Maybe Text

The ID of the group to grant access to, for example, the Active Directory group or identity provider (IdP) group. Required if AuthorizeAllGroups is false or not specified.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

description :: Maybe Text

A brief description of the authorization rule.

authorizeAllGroups :: Maybe Bool

Indicates whether to grant access to all clients. Specify true to grant all clients who successfully establish a VPN connection access to the network. Must be set to true if AccessGroupId is not specified.

clientToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to ensure idempotency.

clientVpnEndpointId :: Text

The ID of the Client VPN endpoint.

targetNetworkCidr :: Text

The IPv4 address range, in CIDR notation, of the network for which access is being authorized.

newAuthorizeClientVpnIngress

Arguments

:: Text
$sel:clientVpnEndpointId:AuthorizeClientVpnIngress'

-> Text
$sel:targetNetworkCidr:AuthorizeClientVpnIngress'

-> AuthorizeClientVpnIngress

Create a value of AuthorizeClientVpnIngress with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accessGroupId:AuthorizeClientVpnIngress', authorizeClientVpnIngress_accessGroupId - The ID of the group to grant access to, for example, the Active Directory group or identity provider (IdP) group. Required if AuthorizeAllGroups is false or not specified.

$sel:dryRun:AuthorizeClientVpnIngress', authorizeClientVpnIngress_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:description:AuthorizeClientVpnIngress', authorizeClientVpnIngress_description - A brief description of the authorization rule.

$sel:authorizeAllGroups:AuthorizeClientVpnIngress', authorizeClientVpnIngress_authorizeAllGroups - Indicates whether to grant access to all clients. Specify true to grant all clients who successfully establish a VPN connection access to the network. Must be set to true if AccessGroupId is not specified.

$sel:clientToken:AuthorizeClientVpnIngress', authorizeClientVpnIngress_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to ensure idempotency.

$sel:clientVpnEndpointId:AuthorizeClientVpnIngress', authorizeClientVpnIngress_clientVpnEndpointId - The ID of the Client VPN endpoint.

$sel:targetNetworkCidr:AuthorizeClientVpnIngress', authorizeClientVpnIngress_targetNetworkCidr - The IPv4 address range, in CIDR notation, of the network for which access is being authorized.

Request Lenses

authorizeClientVpnIngress_accessGroupId :: Lens' AuthorizeClientVpnIngress (Maybe Text)

The ID of the group to grant access to, for example, the Active Directory group or identity provider (IdP) group. Required if AuthorizeAllGroups is false or not specified.

authorizeClientVpnIngress_dryRun :: Lens' AuthorizeClientVpnIngress (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

authorizeClientVpnIngress_description :: Lens' AuthorizeClientVpnIngress (Maybe Text)

A brief description of the authorization rule.

authorizeClientVpnIngress_authorizeAllGroups :: Lens' AuthorizeClientVpnIngress (Maybe Bool)

Indicates whether to grant access to all clients. Specify true to grant all clients who successfully establish a VPN connection access to the network. Must be set to true if AccessGroupId is not specified.

authorizeClientVpnIngress_clientToken :: Lens' AuthorizeClientVpnIngress (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to ensure idempotency.

authorizeClientVpnIngress_clientVpnEndpointId :: Lens' AuthorizeClientVpnIngress Text

The ID of the Client VPN endpoint.

authorizeClientVpnIngress_targetNetworkCidr :: Lens' AuthorizeClientVpnIngress Text

The IPv4 address range, in CIDR notation, of the network for which access is being authorized.

Destructuring the Response

data AuthorizeClientVpnIngressResponse

See: newAuthorizeClientVpnIngressResponse smart constructor.

Constructors

AuthorizeClientVpnIngressResponse'

Fields

status :: Maybe ClientVpnAuthorizationRuleStatus

The current state of the authorization rule.

httpStatus :: Int

The response's http status code.

newAuthorizeClientVpnIngressResponse

Arguments

:: Int
$sel:httpStatus:AuthorizeClientVpnIngressResponse'

-> AuthorizeClientVpnIngressResponse

Create a value of AuthorizeClientVpnIngressResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:AuthorizeClientVpnIngressResponse', authorizeClientVpnIngressResponse_status - The current state of the authorization rule.

$sel:httpStatus:AuthorizeClientVpnIngressResponse', authorizeClientVpnIngressResponse_httpStatus - The response's http status code.

Response Lenses

authorizeClientVpnIngressResponse_status :: Lens' AuthorizeClientVpnIngressResponse (Maybe ClientVpnAuthorizationRuleStatus)

The current state of the authorization rule.

authorizeClientVpnIngressResponse_httpStatus :: Lens' AuthorizeClientVpnIngressResponse Int

The response's http status code.