Network.AWS.EC2.CancelCapacityReservation

Description

Cancels the specified Capacity Reservation, releases the reserved capacity, and changes the Capacity Reservation's state to cancelled.

Instances running in the reserved capacity continue running until you stop them. Stopped instances that target the Capacity Reservation can no longer launch. Modify these instances to either target a different Capacity Reservation, launch On-Demand Instance capacity, or run in any open Capacity Reservation that has matching attributes and sufficient capacity.

Creating a Request

data CancelCapacityReservation

See: newCancelCapacityReservation smart constructor.

Constructors

CancelCapacityReservation'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

capacityReservationId :: Text

The ID of the Capacity Reservation to be cancelled.

newCancelCapacityReservation

Arguments

:: Text
$sel:capacityReservationId:CancelCapacityReservation'

-> CancelCapacityReservation

Create a value of CancelCapacityReservation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:CancelCapacityReservation', cancelCapacityReservation_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:capacityReservationId:CancelCapacityReservation', cancelCapacityReservation_capacityReservationId - The ID of the Capacity Reservation to be cancelled.

Request Lenses

cancelCapacityReservation_dryRun :: Lens' CancelCapacityReservation (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

cancelCapacityReservation_capacityReservationId :: Lens' CancelCapacityReservation Text

The ID of the Capacity Reservation to be cancelled.

Destructuring the Response

data CancelCapacityReservationResponse

See: newCancelCapacityReservationResponse smart constructor.

Constructors

CancelCapacityReservationResponse'

Fields

return' :: Maybe Bool

Returns true if the request succeeds; otherwise, it returns an error.

httpStatus :: Int

The response's http status code.

newCancelCapacityReservationResponse

Arguments

:: Int
$sel:httpStatus:CancelCapacityReservationResponse'

-> CancelCapacityReservationResponse

Create a value of CancelCapacityReservationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:return':CancelCapacityReservationResponse', cancelCapacityReservationResponse_return - Returns true if the request succeeds; otherwise, it returns an error.

$sel:httpStatus:CancelCapacityReservationResponse', cancelCapacityReservationResponse_httpStatus - The response's http status code.

Response Lenses

cancelCapacityReservationResponse_return :: Lens' CancelCapacityReservationResponse (Maybe Bool)

Returns true if the request succeeds; otherwise, it returns an error.

cancelCapacityReservationResponse_httpStatus :: Lens' CancelCapacityReservationResponse Int

The response's http status code.