Network.AWS.EC2.CancelConversionTask

Description

Cancels an active conversion task. The task can be the import of an instance or volume. The action removes all artifacts of the conversion, including a partially uploaded volume or instance. If the conversion is complete or is in the process of transferring the final disk image, the command fails and returns an exception.

For more information, see Importing a Virtual Machine Using the Amazon EC2 CLI.

Creating a Request

data CancelConversionTask

See: newCancelConversionTask smart constructor.

Constructors

CancelConversionTask'

Fields

reasonMessage :: Maybe Text

The reason for canceling the conversion task.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

conversionTaskId :: Text

The ID of the conversion task.

newCancelConversionTask

Arguments

:: Text
$sel:conversionTaskId:CancelConversionTask'

-> CancelConversionTask

Create a value of CancelConversionTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reasonMessage:CancelConversionTask', cancelConversionTask_reasonMessage - The reason for canceling the conversion task.

$sel:dryRun:CancelConversionTask', cancelConversionTask_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:conversionTaskId:CancelConversionTask', cancelConversionTask_conversionTaskId - The ID of the conversion task.

Request Lenses

cancelConversionTask_reasonMessage :: Lens' CancelConversionTask (Maybe Text)

The reason for canceling the conversion task.

cancelConversionTask_dryRun :: Lens' CancelConversionTask (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

cancelConversionTask_conversionTaskId :: Lens' CancelConversionTask Text

The ID of the conversion task.

Destructuring the Response

data CancelConversionTaskResponse

See: newCancelConversionTaskResponse smart constructor.

Constructors

CancelConversionTaskResponse'

newCancelConversionTaskResponse :: CancelConversionTaskResponse

Create a value of CancelConversionTaskResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.