Network.AWS.EC2.CancelReservedInstancesListing

Description

Cancels the specified Reserved Instance listing in the Reserved Instance Marketplace.

For more information, see Reserved Instance Marketplace in the Amazon EC2 User Guide.

Creating a Request

data CancelReservedInstancesListing

Contains the parameters for CancelReservedInstancesListing.

See: newCancelReservedInstancesListing smart constructor.

Constructors

CancelReservedInstancesListing'

Fields

reservedInstancesListingId :: Text

The ID of the Reserved Instance listing.

newCancelReservedInstancesListing

Arguments

:: Text
$sel:reservedInstancesListingId:CancelReservedInstancesListing'

-> CancelReservedInstancesListing

Create a value of CancelReservedInstancesListing with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reservedInstancesListingId:CancelReservedInstancesListing', cancelReservedInstancesListing_reservedInstancesListingId - The ID of the Reserved Instance listing.

Request Lenses

cancelReservedInstancesListing_reservedInstancesListingId :: Lens' CancelReservedInstancesListing Text

The ID of the Reserved Instance listing.

Destructuring the Response

data CancelReservedInstancesListingResponse

Contains the output of CancelReservedInstancesListing.

See: newCancelReservedInstancesListingResponse smart constructor.

Constructors

CancelReservedInstancesListingResponse'

Fields

reservedInstancesListings :: Maybe [ReservedInstancesListing]

The Reserved Instance listing.

httpStatus :: Int

The response's http status code.

newCancelReservedInstancesListingResponse

Arguments

:: Int
$sel:httpStatus:CancelReservedInstancesListingResponse'

-> CancelReservedInstancesListingResponse

Create a value of CancelReservedInstancesListingResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reservedInstancesListings:CancelReservedInstancesListingResponse', cancelReservedInstancesListingResponse_reservedInstancesListings - The Reserved Instance listing.

$sel:httpStatus:CancelReservedInstancesListingResponse', cancelReservedInstancesListingResponse_httpStatus - The response's http status code.

Response Lenses

cancelReservedInstancesListingResponse_reservedInstancesListings :: Lens' CancelReservedInstancesListingResponse (Maybe [ReservedInstancesListing])

The Reserved Instance listing.

cancelReservedInstancesListingResponse_httpStatus :: Lens' CancelReservedInstancesListingResponse Int

The response's http status code.