Network.AWS.EC2.CancelSpotFleetRequests

Description

Cancels the specified Spot Fleet requests.

After you cancel a Spot Fleet request, the Spot Fleet launches no new Spot Instances. You must specify whether the Spot Fleet should also terminate its Spot Instances. If you terminate the instances, the Spot Fleet request enters the cancelled_terminating state. Otherwise, the Spot Fleet request enters the cancelled_running state and the instances continue to run until they are interrupted or you terminate them manually.

Creating a Request

data CancelSpotFleetRequests

Contains the parameters for CancelSpotFleetRequests.

See: newCancelSpotFleetRequests smart constructor.

Constructors

CancelSpotFleetRequests'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

spotFleetRequestIds :: [Text]

The IDs of the Spot Fleet requests.

terminateInstances :: Bool

Indicates whether to terminate instances for a Spot Fleet request if it is canceled successfully.

newCancelSpotFleetRequests

Arguments

:: Bool
$sel:terminateInstances:CancelSpotFleetRequests'

-> CancelSpotFleetRequests

Create a value of CancelSpotFleetRequests with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:CancelSpotFleetRequests', cancelSpotFleetRequests_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:spotFleetRequestIds:CancelSpotFleetRequests', cancelSpotFleetRequests_spotFleetRequestIds - The IDs of the Spot Fleet requests.

$sel:terminateInstances:CancelSpotFleetRequests', cancelSpotFleetRequests_terminateInstances - Indicates whether to terminate instances for a Spot Fleet request if it is canceled successfully.

Request Lenses

cancelSpotFleetRequests_dryRun :: Lens' CancelSpotFleetRequests (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

cancelSpotFleetRequests_spotFleetRequestIds :: Lens' CancelSpotFleetRequests [Text]

The IDs of the Spot Fleet requests.

cancelSpotFleetRequests_terminateInstances :: Lens' CancelSpotFleetRequests Bool

Indicates whether to terminate instances for a Spot Fleet request if it is canceled successfully.

Destructuring the Response

data CancelSpotFleetRequestsResponse

Contains the output of CancelSpotFleetRequests.

See: newCancelSpotFleetRequestsResponse smart constructor.

Constructors

CancelSpotFleetRequestsResponse'

Fields

unsuccessfulFleetRequests :: Maybe [CancelSpotFleetRequestsErrorItem]

Information about the Spot Fleet requests that are not successfully canceled.

successfulFleetRequests :: Maybe [CancelSpotFleetRequestsSuccessItem]

Information about the Spot Fleet requests that are successfully canceled.

httpStatus :: Int

The response's http status code.

newCancelSpotFleetRequestsResponse

Arguments

:: Int
$sel:httpStatus:CancelSpotFleetRequestsResponse'

-> CancelSpotFleetRequestsResponse

Create a value of CancelSpotFleetRequestsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:unsuccessfulFleetRequests:CancelSpotFleetRequestsResponse', cancelSpotFleetRequestsResponse_unsuccessfulFleetRequests - Information about the Spot Fleet requests that are not successfully canceled.

$sel:successfulFleetRequests:CancelSpotFleetRequestsResponse', cancelSpotFleetRequestsResponse_successfulFleetRequests - Information about the Spot Fleet requests that are successfully canceled.

$sel:httpStatus:CancelSpotFleetRequestsResponse', cancelSpotFleetRequestsResponse_httpStatus - The response's http status code.

Response Lenses

cancelSpotFleetRequestsResponse_unsuccessfulFleetRequests :: Lens' CancelSpotFleetRequestsResponse (Maybe [CancelSpotFleetRequestsErrorItem])

Information about the Spot Fleet requests that are not successfully canceled.

cancelSpotFleetRequestsResponse_successfulFleetRequests :: Lens' CancelSpotFleetRequestsResponse (Maybe [CancelSpotFleetRequestsSuccessItem])

Information about the Spot Fleet requests that are successfully canceled.

cancelSpotFleetRequestsResponse_httpStatus :: Lens' CancelSpotFleetRequestsResponse Int

The response's http status code.