Network.AWS.EC2.CancelSpotInstanceRequests

Description

Cancels one or more Spot Instance requests.

Canceling a Spot Instance request does not terminate running Spot Instances associated with the request.

Creating a Request

data CancelSpotInstanceRequests

Contains the parameters for CancelSpotInstanceRequests.

See: newCancelSpotInstanceRequests smart constructor.

Constructors

CancelSpotInstanceRequests'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

spotInstanceRequestIds :: [Text]

One or more Spot Instance request IDs.

newCancelSpotInstanceRequests :: CancelSpotInstanceRequests

Create a value of CancelSpotInstanceRequests with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:CancelSpotInstanceRequests', cancelSpotInstanceRequests_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:spotInstanceRequestIds:CancelSpotInstanceRequests', cancelSpotInstanceRequests_spotInstanceRequestIds - One or more Spot Instance request IDs.

Request Lenses

cancelSpotInstanceRequests_dryRun :: Lens' CancelSpotInstanceRequests (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

cancelSpotInstanceRequests_spotInstanceRequestIds :: Lens' CancelSpotInstanceRequests [Text]

One or more Spot Instance request IDs.

Destructuring the Response

data CancelSpotInstanceRequestsResponse

Contains the output of CancelSpotInstanceRequests.

See: newCancelSpotInstanceRequestsResponse smart constructor.

Constructors

CancelSpotInstanceRequestsResponse'

Fields

cancelledSpotInstanceRequests :: Maybe [CancelledSpotInstanceRequest]

One or more Spot Instance requests.

httpStatus :: Int

The response's http status code.

newCancelSpotInstanceRequestsResponse

Arguments

:: Int
$sel:httpStatus:CancelSpotInstanceRequestsResponse'

-> CancelSpotInstanceRequestsResponse

Create a value of CancelSpotInstanceRequestsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cancelledSpotInstanceRequests:CancelSpotInstanceRequestsResponse', cancelSpotInstanceRequestsResponse_cancelledSpotInstanceRequests - One or more Spot Instance requests.

$sel:httpStatus:CancelSpotInstanceRequestsResponse', cancelSpotInstanceRequestsResponse_httpStatus - The response's http status code.

Response Lenses

cancelSpotInstanceRequestsResponse_cancelledSpotInstanceRequests :: Lens' CancelSpotInstanceRequestsResponse (Maybe [CancelledSpotInstanceRequest])

One or more Spot Instance requests.

cancelSpotInstanceRequestsResponse_httpStatus :: Lens' CancelSpotInstanceRequestsResponse Int

The response's http status code.