Network.AWS.EC2.ConfirmProductInstance

Description

Determines whether a product code is associated with an instance. This action can only be used by the owner of the product code. It is useful when a product code owner must verify whether another user's instance is eligible for support.

Creating a Request

data ConfirmProductInstance

See: newConfirmProductInstance smart constructor.

Constructors

ConfirmProductInstance'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

instanceId :: Text

The ID of the instance.

productCode :: Text

The product code. This must be a product code that you own.

newConfirmProductInstance

Arguments

:: Text
$sel:instanceId:ConfirmProductInstance'

-> Text
$sel:productCode:ConfirmProductInstance'

-> ConfirmProductInstance

Create a value of ConfirmProductInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:ConfirmProductInstance', confirmProductInstance_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:instanceId:ConfirmProductInstance', confirmProductInstance_instanceId - The ID of the instance.

$sel:productCode:ConfirmProductInstance', confirmProductInstance_productCode - The product code. This must be a product code that you own.

Request Lenses

confirmProductInstance_dryRun :: Lens' ConfirmProductInstance (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

confirmProductInstance_instanceId :: Lens' ConfirmProductInstance Text

The ID of the instance.

confirmProductInstance_productCode :: Lens' ConfirmProductInstance Text

The product code. This must be a product code that you own.

Destructuring the Response

data ConfirmProductInstanceResponse

See: newConfirmProductInstanceResponse smart constructor.

Constructors

ConfirmProductInstanceResponse'

Fields

ownerId :: Maybe Text

The Amazon Web Services account ID of the instance owner. This is only present if the product code is attached to the instance.

return' :: Maybe Bool

The return value of the request. Returns true if the specified product code is owned by the requester and associated with the specified instance.

httpStatus :: Int

The response's http status code.

newConfirmProductInstanceResponse

Arguments

:: Int
$sel:httpStatus:ConfirmProductInstanceResponse'

-> ConfirmProductInstanceResponse

Create a value of ConfirmProductInstanceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ownerId:ConfirmProductInstanceResponse', confirmProductInstanceResponse_ownerId - The Amazon Web Services account ID of the instance owner. This is only present if the product code is attached to the instance.

$sel:return':ConfirmProductInstanceResponse', confirmProductInstanceResponse_return - The return value of the request. Returns true if the specified product code is owned by the requester and associated with the specified instance.

$sel:httpStatus:ConfirmProductInstanceResponse', confirmProductInstanceResponse_httpStatus - The response's http status code.

Response Lenses

confirmProductInstanceResponse_ownerId :: Lens' ConfirmProductInstanceResponse (Maybe Text)

The Amazon Web Services account ID of the instance owner. This is only present if the product code is attached to the instance.

confirmProductInstanceResponse_return :: Lens' ConfirmProductInstanceResponse (Maybe Bool)

The return value of the request. Returns true if the specified product code is owned by the requester and associated with the specified instance.

confirmProductInstanceResponse_httpStatus :: Lens' ConfirmProductInstanceResponse Int

The response's http status code.