Network.AWS.EC2.CopyFpgaImage

Description

Copies the specified Amazon FPGA Image (AFI) to the current Region.

Creating a Request

data CopyFpgaImage

See: newCopyFpgaImage smart constructor.

Constructors

CopyFpgaImage'

Fields

clientToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see Ensuring idempotency.

name :: Maybe Text

The name for the new AFI. The default is the name of the source AFI.

description :: Maybe Text

The description for the new AFI.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

sourceFpgaImageId :: Text

The ID of the source AFI.

sourceRegion :: Text

The Region that contains the source AFI.

newCopyFpgaImage

Arguments

:: Text
$sel:sourceFpgaImageId:CopyFpgaImage'

-> Text
$sel:sourceRegion:CopyFpgaImage'

-> CopyFpgaImage

Create a value of CopyFpgaImage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CopyFpgaImage', copyFpgaImage_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see Ensuring idempotency.

$sel:name:CopyFpgaImage', copyFpgaImage_name - The name for the new AFI. The default is the name of the source AFI.

$sel:description:CopyFpgaImage', copyFpgaImage_description - The description for the new AFI.

$sel:dryRun:CopyFpgaImage', copyFpgaImage_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:sourceFpgaImageId:CopyFpgaImage', copyFpgaImage_sourceFpgaImageId - The ID of the source AFI.

$sel:sourceRegion:CopyFpgaImage', copyFpgaImage_sourceRegion - The Region that contains the source AFI.

Request Lenses

copyFpgaImage_clientToken :: Lens' CopyFpgaImage (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see Ensuring idempotency.

copyFpgaImage_name :: Lens' CopyFpgaImage (Maybe Text)

The name for the new AFI. The default is the name of the source AFI.

copyFpgaImage_description :: Lens' CopyFpgaImage (Maybe Text)

The description for the new AFI.

copyFpgaImage_dryRun :: Lens' CopyFpgaImage (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

copyFpgaImage_sourceFpgaImageId :: Lens' CopyFpgaImage Text

The ID of the source AFI.

copyFpgaImage_sourceRegion :: Lens' CopyFpgaImage Text

The Region that contains the source AFI.

Destructuring the Response

data CopyFpgaImageResponse

See: newCopyFpgaImageResponse smart constructor.

Constructors

CopyFpgaImageResponse'

Fields

fpgaImageId :: Maybe Text

The ID of the new AFI.

httpStatus :: Int

The response's http status code.

newCopyFpgaImageResponse

Arguments

:: Int
$sel:httpStatus:CopyFpgaImageResponse'

-> CopyFpgaImageResponse

Create a value of CopyFpgaImageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fpgaImageId:CopyFpgaImageResponse', copyFpgaImageResponse_fpgaImageId - The ID of the new AFI.

$sel:httpStatus:CopyFpgaImageResponse', copyFpgaImageResponse_httpStatus - The response's http status code.

Response Lenses

copyFpgaImageResponse_fpgaImageId :: Lens' CopyFpgaImageResponse (Maybe Text)

The ID of the new AFI.

copyFpgaImageResponse_httpStatus :: Lens' CopyFpgaImageResponse Int

The response's http status code.