Network.AWS.EC2.CreateCapacityReservation

Description

Creates a new Capacity Reservation with the specified attributes.

Capacity Reservations enable you to reserve capacity for your Amazon EC2 instances in a specific Availability Zone for any duration. This gives you the flexibility to selectively add capacity reservations and still get the Regional RI discounts for that usage. By creating Capacity Reservations, you ensure that you always have access to Amazon EC2 capacity when you need it, for as long as you need it. For more information, see Capacity Reservations in the Amazon EC2 User Guide.

Your request to create a Capacity Reservation could fail if Amazon EC2 does not have sufficient capacity to fulfill the request. If your request fails due to Amazon EC2 capacity constraints, either try again at a later time, try in a different Availability Zone, or request a smaller capacity reservation. If your application is flexible across instance types and sizes, try to create a Capacity Reservation with different instance attributes.

Your request could also fail if the requested quantity exceeds your On-Demand Instance limit for the selected instance type. If your request fails due to limit constraints, increase your On-Demand Instance limit for the required instance type and try again. For more information about increasing your instance limits, see Amazon EC2 Service Quotas in the Amazon EC2 User Guide.

Creating a Request

data CreateCapacityReservation

See: newCreateCapacityReservation smart constructor.

Constructors

CreateCapacityReservation'

Fields

tagSpecifications :: Maybe [TagSpecification]

The tags to apply to the Capacity Reservation during launch.

ebsOptimized :: Maybe Bool

Indicates whether the Capacity Reservation supports EBS-optimized instances. This optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal I/O performance. This optimization isn't available with all instance types. Additional usage charges apply when using an EBS- optimized instance.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

outpostArn :: Maybe Text

The Amazon Resource Name (ARN) of the Outpost on which to create the Capacity Reservation.

endDateType :: Maybe EndDateType

Indicates the way in which the Capacity Reservation ends. A Capacity Reservation can have one of the following end types:

unlimited - The Capacity Reservation remains active until you explicitly cancel it. Do not provide an EndDate if the EndDateType is unlimited.

limited - The Capacity Reservation expires automatically at a specified date and time. You must provide an EndDate value if the EndDateType value is limited.

availabilityZoneId :: Maybe Text

The ID of the Availability Zone in which to create the Capacity Reservation.

tenancy :: Maybe CapacityReservationTenancy

Indicates the tenancy of the Capacity Reservation. A Capacity Reservation can have one of the following tenancy settings:

default - The Capacity Reservation is created on hardware that is shared with other Amazon Web Services accounts.

dedicated - The Capacity Reservation is created on single-tenant hardware that is dedicated to a single Amazon Web Services account.

availabilityZone :: Maybe Text

The Availability Zone in which to create the Capacity Reservation.

instanceMatchCriteria :: Maybe InstanceMatchCriteria

Indicates the type of instance launches that the Capacity Reservation accepts. The options include:

open - The Capacity Reservation automatically matches all instances that have matching attributes (instance type, platform, and Availability Zone). Instances that have matching attributes run in the Capacity Reservation automatically without specifying any additional parameters.

targeted - The Capacity Reservation only accepts instances that have matching attributes (instance type, platform, and Availability Zone), and explicitly target the Capacity Reservation. This ensures that only permitted instances can use the reserved capacity.

Default: open

ephemeralStorage :: Maybe Bool

Indicates whether the Capacity Reservation supports instances with temporary, block-level storage.

endDate :: Maybe ISO8601

The date and time at which the Capacity Reservation expires. When a Capacity Reservation expires, the reserved capacity is released and you can no longer launch instances into it. The Capacity Reservation's state changes to expired when it reaches its end date and time.

You must provide an EndDate value if EndDateType is limited. Omit EndDate if EndDateType is unlimited.

If the EndDateType is limited, the Capacity Reservation is cancelled within an hour from the specified time. For example, if you specify 5/31/2019, 13:30:55, the Capacity Reservation is guaranteed to end between 13:30:55 and 14:30:55 on 5/31/2019.

clientToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see Ensure Idempotency.

instanceType :: Text

The instance type for which to reserve capacity. For more information, see Instance types in the Amazon EC2 User Guide.

instancePlatform :: CapacityReservationInstancePlatform

The type of operating system for which to reserve capacity.

instanceCount :: Int

The number of instances for which to reserve capacity.

Valid range: 1 - 1000

newCreateCapacityReservation

Arguments

:: Text
$sel:instanceType:CreateCapacityReservation'

-> CapacityReservationInstancePlatform
$sel:instancePlatform:CreateCapacityReservation'

-> Int
$sel:instanceCount:CreateCapacityReservation'

-> CreateCapacityReservation

Create a value of CreateCapacityReservation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tagSpecifications:CreateCapacityReservation', createCapacityReservation_tagSpecifications - The tags to apply to the Capacity Reservation during launch.

$sel:ebsOptimized:CreateCapacityReservation', createCapacityReservation_ebsOptimized - Indicates whether the Capacity Reservation supports EBS-optimized instances. This optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal I/O performance. This optimization isn't available with all instance types. Additional usage charges apply when using an EBS- optimized instance.

$sel:dryRun:CreateCapacityReservation', createCapacityReservation_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:outpostArn:CreateCapacityReservation', createCapacityReservation_outpostArn - The Amazon Resource Name (ARN) of the Outpost on which to create the Capacity Reservation.

$sel:endDateType:CreateCapacityReservation', createCapacityReservation_endDateType - Indicates the way in which the Capacity Reservation ends. A Capacity Reservation can have one of the following end types:

unlimited - The Capacity Reservation remains active until you explicitly cancel it. Do not provide an EndDate if the EndDateType is unlimited.

limited - The Capacity Reservation expires automatically at a specified date and time. You must provide an EndDate value if the EndDateType value is limited.

$sel:availabilityZoneId:CreateCapacityReservation', createCapacityReservation_availabilityZoneId - The ID of the Availability Zone in which to create the Capacity Reservation.

$sel:tenancy:CreateCapacityReservation', createCapacityReservation_tenancy - Indicates the tenancy of the Capacity Reservation. A Capacity Reservation can have one of the following tenancy settings:

default - The Capacity Reservation is created on hardware that is shared with other Amazon Web Services accounts.

dedicated - The Capacity Reservation is created on single-tenant hardware that is dedicated to a single Amazon Web Services account.

$sel:availabilityZone:CreateCapacityReservation', createCapacityReservation_availabilityZone - The Availability Zone in which to create the Capacity Reservation.

$sel:instanceMatchCriteria:CreateCapacityReservation', createCapacityReservation_instanceMatchCriteria - Indicates the type of instance launches that the Capacity Reservation accepts. The options include:

open - The Capacity Reservation automatically matches all instances that have matching attributes (instance type, platform, and Availability Zone). Instances that have matching attributes run in the Capacity Reservation automatically without specifying any additional parameters.

targeted - The Capacity Reservation only accepts instances that have matching attributes (instance type, platform, and Availability Zone), and explicitly target the Capacity Reservation. This ensures that only permitted instances can use the reserved capacity.

Default: open

$sel:ephemeralStorage:CreateCapacityReservation', createCapacityReservation_ephemeralStorage - Indicates whether the Capacity Reservation supports instances with temporary, block-level storage.

$sel:endDate:CreateCapacityReservation', createCapacityReservation_endDate - The date and time at which the Capacity Reservation expires. When a Capacity Reservation expires, the reserved capacity is released and you can no longer launch instances into it. The Capacity Reservation's state changes to expired when it reaches its end date and time.

You must provide an EndDate value if EndDateType is limited. Omit EndDate if EndDateType is unlimited.

If the EndDateType is limited, the Capacity Reservation is cancelled within an hour from the specified time. For example, if you specify 5/31/2019, 13:30:55, the Capacity Reservation is guaranteed to end between 13:30:55 and 14:30:55 on 5/31/2019.

$sel:clientToken:CreateCapacityReservation', createCapacityReservation_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see Ensure Idempotency.

$sel:instanceType:CreateCapacityReservation', createCapacityReservation_instanceType - The instance type for which to reserve capacity. For more information, see Instance types in the Amazon EC2 User Guide.

$sel:instancePlatform:CreateCapacityReservation', createCapacityReservation_instancePlatform - The type of operating system for which to reserve capacity.

$sel:instanceCount:CreateCapacityReservation', createCapacityReservation_instanceCount - The number of instances for which to reserve capacity.

Valid range: 1 - 1000

Request Lenses

createCapacityReservation_tagSpecifications :: Lens' CreateCapacityReservation (Maybe [TagSpecification])

The tags to apply to the Capacity Reservation during launch.

createCapacityReservation_ebsOptimized :: Lens' CreateCapacityReservation (Maybe Bool)

Indicates whether the Capacity Reservation supports EBS-optimized instances. This optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal I/O performance. This optimization isn't available with all instance types. Additional usage charges apply when using an EBS- optimized instance.

createCapacityReservation_dryRun :: Lens' CreateCapacityReservation (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

createCapacityReservation_outpostArn :: Lens' CreateCapacityReservation (Maybe Text)

The Amazon Resource Name (ARN) of the Outpost on which to create the Capacity Reservation.

createCapacityReservation_endDateType :: Lens' CreateCapacityReservation (Maybe EndDateType)

Indicates the way in which the Capacity Reservation ends. A Capacity Reservation can have one of the following end types:

unlimited - The Capacity Reservation remains active until you explicitly cancel it. Do not provide an EndDate if the EndDateType is unlimited.

limited - The Capacity Reservation expires automatically at a specified date and time. You must provide an EndDate value if the EndDateType value is limited.

createCapacityReservation_availabilityZoneId :: Lens' CreateCapacityReservation (Maybe Text)

The ID of the Availability Zone in which to create the Capacity Reservation.

createCapacityReservation_tenancy :: Lens' CreateCapacityReservation (Maybe CapacityReservationTenancy)

Indicates the tenancy of the Capacity Reservation. A Capacity Reservation can have one of the following tenancy settings:

default - The Capacity Reservation is created on hardware that is shared with other Amazon Web Services accounts.

dedicated - The Capacity Reservation is created on single-tenant hardware that is dedicated to a single Amazon Web Services account.

createCapacityReservation_availabilityZone :: Lens' CreateCapacityReservation (Maybe Text)

The Availability Zone in which to create the Capacity Reservation.

createCapacityReservation_instanceMatchCriteria :: Lens' CreateCapacityReservation (Maybe InstanceMatchCriteria)

Indicates the type of instance launches that the Capacity Reservation accepts. The options include:

open - The Capacity Reservation automatically matches all instances that have matching attributes (instance type, platform, and Availability Zone). Instances that have matching attributes run in the Capacity Reservation automatically without specifying any additional parameters.

targeted - The Capacity Reservation only accepts instances that have matching attributes (instance type, platform, and Availability Zone), and explicitly target the Capacity Reservation. This ensures that only permitted instances can use the reserved capacity.

Default: open

createCapacityReservation_ephemeralStorage :: Lens' CreateCapacityReservation (Maybe Bool)

Indicates whether the Capacity Reservation supports instances with temporary, block-level storage.

createCapacityReservation_endDate :: Lens' CreateCapacityReservation (Maybe UTCTime)

The date and time at which the Capacity Reservation expires. When a Capacity Reservation expires, the reserved capacity is released and you can no longer launch instances into it. The Capacity Reservation's state changes to expired when it reaches its end date and time.

You must provide an EndDate value if EndDateType is limited. Omit EndDate if EndDateType is unlimited.

If the EndDateType is limited, the Capacity Reservation is cancelled within an hour from the specified time. For example, if you specify 5/31/2019, 13:30:55, the Capacity Reservation is guaranteed to end between 13:30:55 and 14:30:55 on 5/31/2019.

createCapacityReservation_clientToken :: Lens' CreateCapacityReservation (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see Ensure Idempotency.

createCapacityReservation_instanceType :: Lens' CreateCapacityReservation Text

The instance type for which to reserve capacity. For more information, see Instance types in the Amazon EC2 User Guide.

createCapacityReservation_instancePlatform :: Lens' CreateCapacityReservation CapacityReservationInstancePlatform

The type of operating system for which to reserve capacity.

createCapacityReservation_instanceCount :: Lens' CreateCapacityReservation Int

The number of instances for which to reserve capacity.

Valid range: 1 - 1000

Destructuring the Response

data CreateCapacityReservationResponse

See: newCreateCapacityReservationResponse smart constructor.

Constructors

CreateCapacityReservationResponse'

Fields

capacityReservation :: Maybe CapacityReservation

Information about the Capacity Reservation.

httpStatus :: Int

The response's http status code.

newCreateCapacityReservationResponse

Arguments

:: Int
$sel:httpStatus:CreateCapacityReservationResponse'

-> CreateCapacityReservationResponse

Create a value of CreateCapacityReservationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:capacityReservation:CreateCapacityReservationResponse', createCapacityReservationResponse_capacityReservation - Information about the Capacity Reservation.

$sel:httpStatus:CreateCapacityReservationResponse', createCapacityReservationResponse_httpStatus - The response's http status code.

Response Lenses

createCapacityReservationResponse_capacityReservation :: Lens' CreateCapacityReservationResponse (Maybe CapacityReservation)

Information about the Capacity Reservation.

createCapacityReservationResponse_httpStatus :: Lens' CreateCapacityReservationResponse Int

The response's http status code.