Network.AWS.EC2.CreateCapacityReservationFleet

Description

Creates a Capacity Reservation Fleet. For more information, see Create a Capacity Reservation Fleet in the Amazon EC2 User Guide.

Creating a Request

data CreateCapacityReservationFleet

See: newCreateCapacityReservationFleet smart constructor.

Constructors

CreateCapacityReservationFleet'

Fields

clientToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see Ensure Idempotency.

endDate :: Maybe ISO8601

The date and time at which the Capacity Reservation Fleet expires. When the Capacity Reservation Fleet expires, its state changes to expired and all of the Capacity Reservations in the Fleet expire.

The Capacity Reservation Fleet expires within an hour after the specified time. For example, if you specify 5/31/2019, 13:30:55, the Capacity Reservation Fleet is guaranteed to expire between 13:30:55 and 14:30:55 on 5/31/2019.

instanceMatchCriteria :: Maybe FleetInstanceMatchCriteria

Indicates the type of instance launches that the Capacity Reservation Fleet accepts. All Capacity Reservations in the Fleet inherit this instance matching criteria.

Currently, Capacity Reservation Fleets support open instance matching criteria only. This means that instances that have matching attributes (instance type, platform, and Availability Zone) run in the Capacity Reservations automatically. Instances do not need to explicitly target a Capacity Reservation Fleet to use its reserved capacity.

tagSpecifications :: Maybe [TagSpecification]

The tags to assign to the Capacity Reservation Fleet. The tags are automatically assigned to the Capacity Reservations in the Fleet.

tenancy :: Maybe FleetCapacityReservationTenancy

Indicates the tenancy of the Capacity Reservation Fleet. All Capacity Reservations in the Fleet inherit this tenancy. The Capacity Reservation Fleet can have one of the following tenancy settings:

default - The Capacity Reservation Fleet is created on hardware that is shared with other Amazon Web Services accounts.

dedicated - The Capacity Reservations are created on single-tenant hardware that is dedicated to a single Amazon Web Services account.

allocationStrategy :: Maybe Text

The strategy used by the Capacity Reservation Fleet to determine which of the specified instance types to use. Currently, only the prioritized allocation strategy is supported. For more information, see Allocation strategy in the Amazon EC2 User Guide.

Valid values: prioritized

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

instanceTypeSpecifications :: [ReservationFleetInstanceSpecification]

Information about the instance types for which to reserve the capacity.

totalTargetCapacity :: Int

The total number of capacity units to be reserved by the Capacity Reservation Fleet. This value, together with the instance type weights that you assign to each instance type used by the Fleet determine the number of instances for which the Fleet reserves capacity. Both values are based on units that make sense for your workload. For more information, see Total target capacity in the Amazon EC2 User Guide.

newCreateCapacityReservationFleet

Arguments

:: Int
$sel:totalTargetCapacity:CreateCapacityReservationFleet'

-> CreateCapacityReservationFleet

Create a value of CreateCapacityReservationFleet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateCapacityReservationFleet', createCapacityReservationFleet_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see Ensure Idempotency.

$sel:endDate:CreateCapacityReservationFleet', createCapacityReservationFleet_endDate - The date and time at which the Capacity Reservation Fleet expires. When the Capacity Reservation Fleet expires, its state changes to expired and all of the Capacity Reservations in the Fleet expire.

The Capacity Reservation Fleet expires within an hour after the specified time. For example, if you specify 5/31/2019, 13:30:55, the Capacity Reservation Fleet is guaranteed to expire between 13:30:55 and 14:30:55 on 5/31/2019.

$sel:instanceMatchCriteria:CreateCapacityReservationFleet', createCapacityReservationFleet_instanceMatchCriteria - Indicates the type of instance launches that the Capacity Reservation Fleet accepts. All Capacity Reservations in the Fleet inherit this instance matching criteria.

Currently, Capacity Reservation Fleets support open instance matching criteria only. This means that instances that have matching attributes (instance type, platform, and Availability Zone) run in the Capacity Reservations automatically. Instances do not need to explicitly target a Capacity Reservation Fleet to use its reserved capacity.

$sel:tagSpecifications:CreateCapacityReservationFleet', createCapacityReservationFleet_tagSpecifications - The tags to assign to the Capacity Reservation Fleet. The tags are automatically assigned to the Capacity Reservations in the Fleet.

$sel:tenancy:CreateCapacityReservationFleet', createCapacityReservationFleet_tenancy - Indicates the tenancy of the Capacity Reservation Fleet. All Capacity Reservations in the Fleet inherit this tenancy. The Capacity Reservation Fleet can have one of the following tenancy settings:

default - The Capacity Reservation Fleet is created on hardware that is shared with other Amazon Web Services accounts.

dedicated - The Capacity Reservations are created on single-tenant hardware that is dedicated to a single Amazon Web Services account.

$sel:allocationStrategy:CreateCapacityReservationFleet', createCapacityReservationFleet_allocationStrategy - The strategy used by the Capacity Reservation Fleet to determine which of the specified instance types to use. Currently, only the prioritized allocation strategy is supported. For more information, see Allocation strategy in the Amazon EC2 User Guide.

Valid values: prioritized

$sel:dryRun:CreateCapacityReservationFleet', createCapacityReservationFleet_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:instanceTypeSpecifications:CreateCapacityReservationFleet', createCapacityReservationFleet_instanceTypeSpecifications - Information about the instance types for which to reserve the capacity.

$sel:totalTargetCapacity:CreateCapacityReservationFleet', createCapacityReservationFleet_totalTargetCapacity - The total number of capacity units to be reserved by the Capacity Reservation Fleet. This value, together with the instance type weights that you assign to each instance type used by the Fleet determine the number of instances for which the Fleet reserves capacity. Both values are based on units that make sense for your workload. For more information, see Total target capacity in the Amazon EC2 User Guide.

Request Lenses

createCapacityReservationFleet_clientToken :: Lens' CreateCapacityReservationFleet (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see Ensure Idempotency.

createCapacityReservationFleet_endDate :: Lens' CreateCapacityReservationFleet (Maybe UTCTime)

The date and time at which the Capacity Reservation Fleet expires. When the Capacity Reservation Fleet expires, its state changes to expired and all of the Capacity Reservations in the Fleet expire.

The Capacity Reservation Fleet expires within an hour after the specified time. For example, if you specify 5/31/2019, 13:30:55, the Capacity Reservation Fleet is guaranteed to expire between 13:30:55 and 14:30:55 on 5/31/2019.

createCapacityReservationFleet_instanceMatchCriteria :: Lens' CreateCapacityReservationFleet (Maybe FleetInstanceMatchCriteria)

Indicates the type of instance launches that the Capacity Reservation Fleet accepts. All Capacity Reservations in the Fleet inherit this instance matching criteria.

Currently, Capacity Reservation Fleets support open instance matching criteria only. This means that instances that have matching attributes (instance type, platform, and Availability Zone) run in the Capacity Reservations automatically. Instances do not need to explicitly target a Capacity Reservation Fleet to use its reserved capacity.

createCapacityReservationFleet_tagSpecifications :: Lens' CreateCapacityReservationFleet (Maybe [TagSpecification])

The tags to assign to the Capacity Reservation Fleet. The tags are automatically assigned to the Capacity Reservations in the Fleet.

createCapacityReservationFleet_tenancy :: Lens' CreateCapacityReservationFleet (Maybe FleetCapacityReservationTenancy)

Indicates the tenancy of the Capacity Reservation Fleet. All Capacity Reservations in the Fleet inherit this tenancy. The Capacity Reservation Fleet can have one of the following tenancy settings:

default - The Capacity Reservation Fleet is created on hardware that is shared with other Amazon Web Services accounts.

dedicated - The Capacity Reservations are created on single-tenant hardware that is dedicated to a single Amazon Web Services account.

createCapacityReservationFleet_allocationStrategy :: Lens' CreateCapacityReservationFleet (Maybe Text)

The strategy used by the Capacity Reservation Fleet to determine which of the specified instance types to use. Currently, only the prioritized allocation strategy is supported. For more information, see Allocation strategy in the Amazon EC2 User Guide.

Valid values: prioritized

createCapacityReservationFleet_dryRun :: Lens' CreateCapacityReservationFleet (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

createCapacityReservationFleet_instanceTypeSpecifications :: Lens' CreateCapacityReservationFleet [ReservationFleetInstanceSpecification]

Information about the instance types for which to reserve the capacity.

createCapacityReservationFleet_totalTargetCapacity :: Lens' CreateCapacityReservationFleet Int

The total number of capacity units to be reserved by the Capacity Reservation Fleet. This value, together with the instance type weights that you assign to each instance type used by the Fleet determine the number of instances for which the Fleet reserves capacity. Both values are based on units that make sense for your workload. For more information, see Total target capacity in the Amazon EC2 User Guide.

Destructuring the Response

data CreateCapacityReservationFleetResponse

See: newCreateCapacityReservationFleetResponse smart constructor.

Constructors

CreateCapacityReservationFleetResponse'

Fields

capacityReservationFleetId :: Maybe Text

The ID of the Capacity Reservation Fleet.

state :: Maybe CapacityReservationFleetState

The status of the Capacity Reservation Fleet.

totalFulfilledCapacity :: Maybe Double

The requested capacity units that have been successfully reserved.

endDate :: Maybe ISO8601

The date and time at which the Capacity Reservation Fleet expires.

instanceMatchCriteria :: Maybe FleetInstanceMatchCriteria

The instance matching criteria for the Capacity Reservation Fleet.

fleetCapacityReservations :: Maybe [FleetCapacityReservation]

Information about the individual Capacity Reservations in the Capacity Reservation Fleet.

totalTargetCapacity :: Maybe Int

The total number of capacity units for which the Capacity Reservation Fleet reserves capacity.

tenancy :: Maybe FleetCapacityReservationTenancy

Indicates the tenancy of Capacity Reservation Fleet.

allocationStrategy :: Maybe Text

The allocation strategy used by the Capacity Reservation Fleet.

createTime :: Maybe ISO8601

The date and time at which the Capacity Reservation Fleet was created.

tags :: Maybe [Tag]

The tags assigned to the Capacity Reservation Fleet.

httpStatus :: Int

The response's http status code.

newCreateCapacityReservationFleetResponse

Arguments

:: Int
$sel:httpStatus:CreateCapacityReservationFleetResponse'

-> CreateCapacityReservationFleetResponse

Create a value of CreateCapacityReservationFleetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:capacityReservationFleetId:CreateCapacityReservationFleetResponse', createCapacityReservationFleetResponse_capacityReservationFleetId - The ID of the Capacity Reservation Fleet.

$sel:state:CreateCapacityReservationFleetResponse', createCapacityReservationFleetResponse_state - The status of the Capacity Reservation Fleet.

$sel:totalFulfilledCapacity:CreateCapacityReservationFleetResponse', createCapacityReservationFleetResponse_totalFulfilledCapacity - The requested capacity units that have been successfully reserved.

$sel:endDate:CreateCapacityReservationFleet', createCapacityReservationFleetResponse_endDate - The date and time at which the Capacity Reservation Fleet expires.

$sel:instanceMatchCriteria:CreateCapacityReservationFleet', createCapacityReservationFleetResponse_instanceMatchCriteria - The instance matching criteria for the Capacity Reservation Fleet.

$sel:fleetCapacityReservations:CreateCapacityReservationFleetResponse', createCapacityReservationFleetResponse_fleetCapacityReservations - Information about the individual Capacity Reservations in the Capacity Reservation Fleet.

$sel:totalTargetCapacity:CreateCapacityReservationFleet', createCapacityReservationFleetResponse_totalTargetCapacity - The total number of capacity units for which the Capacity Reservation Fleet reserves capacity.

$sel:tenancy:CreateCapacityReservationFleet', createCapacityReservationFleetResponse_tenancy - Indicates the tenancy of Capacity Reservation Fleet.

$sel:allocationStrategy:CreateCapacityReservationFleet', createCapacityReservationFleetResponse_allocationStrategy - The allocation strategy used by the Capacity Reservation Fleet.

$sel:createTime:CreateCapacityReservationFleetResponse', createCapacityReservationFleetResponse_createTime - The date and time at which the Capacity Reservation Fleet was created.

$sel:tags:CreateCapacityReservationFleetResponse', createCapacityReservationFleetResponse_tags - The tags assigned to the Capacity Reservation Fleet.

$sel:httpStatus:CreateCapacityReservationFleetResponse', createCapacityReservationFleetResponse_httpStatus - The response's http status code.

Response Lenses

createCapacityReservationFleetResponse_capacityReservationFleetId :: Lens' CreateCapacityReservationFleetResponse (Maybe Text)

The ID of the Capacity Reservation Fleet.

createCapacityReservationFleetResponse_state :: Lens' CreateCapacityReservationFleetResponse (Maybe CapacityReservationFleetState)

The status of the Capacity Reservation Fleet.

createCapacityReservationFleetResponse_totalFulfilledCapacity :: Lens' CreateCapacityReservationFleetResponse (Maybe Double)

The requested capacity units that have been successfully reserved.

createCapacityReservationFleetResponse_endDate :: Lens' CreateCapacityReservationFleetResponse (Maybe UTCTime)

The date and time at which the Capacity Reservation Fleet expires.

createCapacityReservationFleetResponse_instanceMatchCriteria :: Lens' CreateCapacityReservationFleetResponse (Maybe FleetInstanceMatchCriteria)

The instance matching criteria for the Capacity Reservation Fleet.

createCapacityReservationFleetResponse_fleetCapacityReservations :: Lens' CreateCapacityReservationFleetResponse (Maybe [FleetCapacityReservation])

Information about the individual Capacity Reservations in the Capacity Reservation Fleet.

createCapacityReservationFleetResponse_totalTargetCapacity :: Lens' CreateCapacityReservationFleetResponse (Maybe Int)

The total number of capacity units for which the Capacity Reservation Fleet reserves capacity.

createCapacityReservationFleetResponse_tenancy :: Lens' CreateCapacityReservationFleetResponse (Maybe FleetCapacityReservationTenancy)

Indicates the tenancy of Capacity Reservation Fleet.

createCapacityReservationFleetResponse_allocationStrategy :: Lens' CreateCapacityReservationFleetResponse (Maybe Text)

The allocation strategy used by the Capacity Reservation Fleet.

createCapacityReservationFleetResponse_createTime :: Lens' CreateCapacityReservationFleetResponse (Maybe UTCTime)

The date and time at which the Capacity Reservation Fleet was created.

createCapacityReservationFleetResponse_tags :: Lens' CreateCapacityReservationFleetResponse (Maybe [Tag])

The tags assigned to the Capacity Reservation Fleet.

createCapacityReservationFleetResponse_httpStatus :: Lens' CreateCapacityReservationFleetResponse Int

The response's http status code.