Network.AWS.EC2.CreateClientVpnEndpoint

Description

Creates a Client VPN endpoint. A Client VPN endpoint is the resource you create and configure to enable and manage client VPN sessions. It is the destination endpoint at which all client VPN sessions are terminated.

Creating a Request

data CreateClientVpnEndpoint

See: newCreateClientVpnEndpoint smart constructor.

Constructors

CreateClientVpnEndpoint'

Fields

securityGroupIds :: Maybe [Text]

The IDs of one or more security groups to apply to the target network. You must also specify the ID of the VPC that contains the security groups.

tagSpecifications :: Maybe [TagSpecification]

The tags to apply to the Client VPN endpoint during creation.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

transportProtocol :: Maybe TransportProtocol

The transport protocol to be used by the VPN session.

Default value: udp

clientConnectOptions :: Maybe ClientConnectOptions

The options for managing connection authorization for new client connections.

dnsServers :: Maybe [Text]

Information about the DNS servers to be used for DNS resolution. A Client VPN endpoint can have up to two DNS servers. If no DNS server is specified, the DNS address configured on the device is used for the DNS server.

vpnPort :: Maybe Int

The port number to assign to the Client VPN endpoint for TCP and UDP traffic.

Valid Values: 443 | 1194

Default Value: 443

description :: Maybe Text

A brief description of the Client VPN endpoint.

vpcId :: Maybe Text

The ID of the VPC to associate with the Client VPN endpoint. If no security group IDs are specified in the request, the default security group for the VPC is applied.

selfServicePortal :: Maybe SelfServicePortal

Specify whether to enable the self-service portal for the Client VPN endpoint.

Default Value: enabled

splitTunnel :: Maybe Bool

Indicates whether split-tunnel is enabled on the Client VPN endpoint.

By default, split-tunnel on a VPN endpoint is disabled.

For information about split-tunnel VPN endpoints, see Split-tunnel Client VPN endpoint in the Client VPN Administrator Guide.

clientToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to ensure idempotency.

clientCidrBlock :: Text

The IPv4 address range, in CIDR notation, from which to assign client IP addresses. The address range cannot overlap with the local CIDR of the VPC in which the associated subnet is located, or the routes that you add manually. The address range cannot be changed after the Client VPN endpoint has been created. The CIDR block should be /22 or greater.

serverCertificateArn :: Text

The ARN of the server certificate. For more information, see the Certificate Manager User Guide.

authenticationOptions :: [ClientVpnAuthenticationRequest]

Information about the authentication method to be used to authenticate clients.

connectionLogOptions :: ConnectionLogOptions

Information about the client connection logging options.

If you enable client connection logging, data about client connections is sent to a Cloudwatch Logs log stream. The following information is logged:

Client connection requests

Client connection results (successful and unsuccessful)

Reasons for unsuccessful client connection requests

Client connection termination time

newCreateClientVpnEndpoint

Arguments

:: Text
$sel:clientCidrBlock:CreateClientVpnEndpoint'

-> Text
$sel:serverCertificateArn:CreateClientVpnEndpoint'

-> ConnectionLogOptions
$sel:connectionLogOptions:CreateClientVpnEndpoint'

-> CreateClientVpnEndpoint

Create a value of CreateClientVpnEndpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:securityGroupIds:CreateClientVpnEndpoint', createClientVpnEndpoint_securityGroupIds - The IDs of one or more security groups to apply to the target network. You must also specify the ID of the VPC that contains the security groups.

$sel:tagSpecifications:CreateClientVpnEndpoint', createClientVpnEndpoint_tagSpecifications - The tags to apply to the Client VPN endpoint during creation.

$sel:dryRun:CreateClientVpnEndpoint', createClientVpnEndpoint_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:transportProtocol:CreateClientVpnEndpoint', createClientVpnEndpoint_transportProtocol - The transport protocol to be used by the VPN session.

Default value: udp

$sel:clientConnectOptions:CreateClientVpnEndpoint', createClientVpnEndpoint_clientConnectOptions - The options for managing connection authorization for new client connections.

$sel:dnsServers:CreateClientVpnEndpoint', createClientVpnEndpoint_dnsServers - Information about the DNS servers to be used for DNS resolution. A Client VPN endpoint can have up to two DNS servers. If no DNS server is specified, the DNS address configured on the device is used for the DNS server.

$sel:vpnPort:CreateClientVpnEndpoint', createClientVpnEndpoint_vpnPort - The port number to assign to the Client VPN endpoint for TCP and UDP traffic.

Valid Values: 443 | 1194

Default Value: 443

$sel:description:CreateClientVpnEndpoint', createClientVpnEndpoint_description - A brief description of the Client VPN endpoint.

$sel:vpcId:CreateClientVpnEndpoint', createClientVpnEndpoint_vpcId - The ID of the VPC to associate with the Client VPN endpoint. If no security group IDs are specified in the request, the default security group for the VPC is applied.

$sel:selfServicePortal:CreateClientVpnEndpoint', createClientVpnEndpoint_selfServicePortal - Specify whether to enable the self-service portal for the Client VPN endpoint.

Default Value: enabled

$sel:splitTunnel:CreateClientVpnEndpoint', createClientVpnEndpoint_splitTunnel - Indicates whether split-tunnel is enabled on the Client VPN endpoint.

By default, split-tunnel on a VPN endpoint is disabled.

For information about split-tunnel VPN endpoints, see Split-tunnel Client VPN endpoint in the Client VPN Administrator Guide.

$sel:clientToken:CreateClientVpnEndpoint', createClientVpnEndpoint_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to ensure idempotency.

$sel:clientCidrBlock:CreateClientVpnEndpoint', createClientVpnEndpoint_clientCidrBlock - The IPv4 address range, in CIDR notation, from which to assign client IP addresses. The address range cannot overlap with the local CIDR of the VPC in which the associated subnet is located, or the routes that you add manually. The address range cannot be changed after the Client VPN endpoint has been created. The CIDR block should be /22 or greater.

$sel:serverCertificateArn:CreateClientVpnEndpoint', createClientVpnEndpoint_serverCertificateArn - The ARN of the server certificate. For more information, see the Certificate Manager User Guide.

$sel:authenticationOptions:CreateClientVpnEndpoint', createClientVpnEndpoint_authenticationOptions - Information about the authentication method to be used to authenticate clients.

$sel:connectionLogOptions:CreateClientVpnEndpoint', createClientVpnEndpoint_connectionLogOptions - Information about the client connection logging options.

If you enable client connection logging, data about client connections is sent to a Cloudwatch Logs log stream. The following information is logged:

Client connection requests

Client connection results (successful and unsuccessful)

Reasons for unsuccessful client connection requests

Client connection termination time

Request Lenses

createClientVpnEndpoint_securityGroupIds :: Lens' CreateClientVpnEndpoint (Maybe [Text])

The IDs of one or more security groups to apply to the target network. You must also specify the ID of the VPC that contains the security groups.

createClientVpnEndpoint_tagSpecifications :: Lens' CreateClientVpnEndpoint (Maybe [TagSpecification])

The tags to apply to the Client VPN endpoint during creation.

createClientVpnEndpoint_dryRun :: Lens' CreateClientVpnEndpoint (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

createClientVpnEndpoint_transportProtocol :: Lens' CreateClientVpnEndpoint (Maybe TransportProtocol)

The transport protocol to be used by the VPN session.

Default value: udp

createClientVpnEndpoint_clientConnectOptions :: Lens' CreateClientVpnEndpoint (Maybe ClientConnectOptions)

The options for managing connection authorization for new client connections.

createClientVpnEndpoint_dnsServers :: Lens' CreateClientVpnEndpoint (Maybe [Text])

Information about the DNS servers to be used for DNS resolution. A Client VPN endpoint can have up to two DNS servers. If no DNS server is specified, the DNS address configured on the device is used for the DNS server.

createClientVpnEndpoint_vpnPort :: Lens' CreateClientVpnEndpoint (Maybe Int)

The port number to assign to the Client VPN endpoint for TCP and UDP traffic.

Valid Values: 443 | 1194

Default Value: 443

createClientVpnEndpoint_description :: Lens' CreateClientVpnEndpoint (Maybe Text)

A brief description of the Client VPN endpoint.

createClientVpnEndpoint_vpcId :: Lens' CreateClientVpnEndpoint (Maybe Text)

The ID of the VPC to associate with the Client VPN endpoint. If no security group IDs are specified in the request, the default security group for the VPC is applied.

createClientVpnEndpoint_selfServicePortal :: Lens' CreateClientVpnEndpoint (Maybe SelfServicePortal)

Specify whether to enable the self-service portal for the Client VPN endpoint.

Default Value: enabled

createClientVpnEndpoint_splitTunnel :: Lens' CreateClientVpnEndpoint (Maybe Bool)

Indicates whether split-tunnel is enabled on the Client VPN endpoint.

By default, split-tunnel on a VPN endpoint is disabled.

For information about split-tunnel VPN endpoints, see Split-tunnel Client VPN endpoint in the Client VPN Administrator Guide.

createClientVpnEndpoint_clientToken :: Lens' CreateClientVpnEndpoint (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to ensure idempotency.

createClientVpnEndpoint_clientCidrBlock :: Lens' CreateClientVpnEndpoint Text

The IPv4 address range, in CIDR notation, from which to assign client IP addresses. The address range cannot overlap with the local CIDR of the VPC in which the associated subnet is located, or the routes that you add manually. The address range cannot be changed after the Client VPN endpoint has been created. The CIDR block should be /22 or greater.

createClientVpnEndpoint_serverCertificateArn :: Lens' CreateClientVpnEndpoint Text

The ARN of the server certificate. For more information, see the Certificate Manager User Guide.

createClientVpnEndpoint_authenticationOptions :: Lens' CreateClientVpnEndpoint [ClientVpnAuthenticationRequest]

Information about the authentication method to be used to authenticate clients.

createClientVpnEndpoint_connectionLogOptions :: Lens' CreateClientVpnEndpoint ConnectionLogOptions

Information about the client connection logging options.

If you enable client connection logging, data about client connections is sent to a Cloudwatch Logs log stream. The following information is logged:

Client connection requests

Client connection results (successful and unsuccessful)

Reasons for unsuccessful client connection requests

Client connection termination time

Destructuring the Response

data CreateClientVpnEndpointResponse

See: newCreateClientVpnEndpointResponse smart constructor.

Constructors

CreateClientVpnEndpointResponse'

Fields

clientVpnEndpointId :: Maybe Text

The ID of the Client VPN endpoint.

status :: Maybe ClientVpnEndpointStatus

The current state of the Client VPN endpoint.

dnsName :: Maybe Text

The DNS name to be used by clients when establishing their VPN session.

httpStatus :: Int

The response's http status code.

newCreateClientVpnEndpointResponse

Arguments

:: Int
$sel:httpStatus:CreateClientVpnEndpointResponse'

-> CreateClientVpnEndpointResponse

Create a value of CreateClientVpnEndpointResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientVpnEndpointId:CreateClientVpnEndpointResponse', createClientVpnEndpointResponse_clientVpnEndpointId - The ID of the Client VPN endpoint.

$sel:status:CreateClientVpnEndpointResponse', createClientVpnEndpointResponse_status - The current state of the Client VPN endpoint.

$sel:dnsName:CreateClientVpnEndpointResponse', createClientVpnEndpointResponse_dnsName - The DNS name to be used by clients when establishing their VPN session.

$sel:httpStatus:CreateClientVpnEndpointResponse', createClientVpnEndpointResponse_httpStatus - The response's http status code.

Response Lenses

createClientVpnEndpointResponse_clientVpnEndpointId :: Lens' CreateClientVpnEndpointResponse (Maybe Text)

The ID of the Client VPN endpoint.

createClientVpnEndpointResponse_status :: Lens' CreateClientVpnEndpointResponse (Maybe ClientVpnEndpointStatus)

The current state of the Client VPN endpoint.

createClientVpnEndpointResponse_dnsName :: Lens' CreateClientVpnEndpointResponse (Maybe Text)

The DNS name to be used by clients when establishing their VPN session.

createClientVpnEndpointResponse_httpStatus :: Lens' CreateClientVpnEndpointResponse Int

The response's http status code.