Network.AWS.EC2.CreateClientVpnRoute

Description

Adds a route to a network to a Client VPN endpoint. Each Client VPN endpoint has a route table that describes the available destination network routes. Each route in the route table specifies the path for traﬃc to speciﬁc resources or networks.

Creating a Request

data CreateClientVpnRoute

See: newCreateClientVpnRoute smart constructor.

Constructors

CreateClientVpnRoute'

Fields

clientToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to ensure idempotency.

description :: Maybe Text

A brief description of the route.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

clientVpnEndpointId :: Text

The ID of the Client VPN endpoint to which to add the route.

destinationCidrBlock :: Text

The IPv4 address range, in CIDR notation, of the route destination. For example:

To add a route for Internet access, enter 0.0.0.0/0

To add a route for a peered VPC, enter the peered VPC's IPv4 CIDR range

To add a route for an on-premises network, enter the Amazon Web Services Site-to-Site VPN connection's IPv4 CIDR range

To add a route for the local network, enter the client CIDR range

targetVpcSubnetId :: Text

The ID of the subnet through which you want to route traffic. The specified subnet must be an existing target network of the Client VPN endpoint.

Alternatively, if you're adding a route for the local network, specify local.

newCreateClientVpnRoute

Arguments

:: Text
$sel:clientVpnEndpointId:CreateClientVpnRoute'

-> Text
$sel:destinationCidrBlock:CreateClientVpnRoute'

-> Text
$sel:targetVpcSubnetId:CreateClientVpnRoute'

-> CreateClientVpnRoute

Create a value of CreateClientVpnRoute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateClientVpnRoute', createClientVpnRoute_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to ensure idempotency.

$sel:description:CreateClientVpnRoute', createClientVpnRoute_description - A brief description of the route.

$sel:dryRun:CreateClientVpnRoute', createClientVpnRoute_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:clientVpnEndpointId:CreateClientVpnRoute', createClientVpnRoute_clientVpnEndpointId - The ID of the Client VPN endpoint to which to add the route.

$sel:destinationCidrBlock:CreateClientVpnRoute', createClientVpnRoute_destinationCidrBlock - The IPv4 address range, in CIDR notation, of the route destination. For example:

To add a route for Internet access, enter 0.0.0.0/0

To add a route for a peered VPC, enter the peered VPC's IPv4 CIDR range

To add a route for an on-premises network, enter the Amazon Web Services Site-to-Site VPN connection's IPv4 CIDR range

To add a route for the local network, enter the client CIDR range

$sel:targetVpcSubnetId:CreateClientVpnRoute', createClientVpnRoute_targetVpcSubnetId - The ID of the subnet through which you want to route traffic. The specified subnet must be an existing target network of the Client VPN endpoint.

Alternatively, if you're adding a route for the local network, specify local.

Request Lenses

createClientVpnRoute_clientToken :: Lens' CreateClientVpnRoute (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to ensure idempotency.

createClientVpnRoute_description :: Lens' CreateClientVpnRoute (Maybe Text)

A brief description of the route.

createClientVpnRoute_dryRun :: Lens' CreateClientVpnRoute (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

createClientVpnRoute_clientVpnEndpointId :: Lens' CreateClientVpnRoute Text

The ID of the Client VPN endpoint to which to add the route.

createClientVpnRoute_destinationCidrBlock :: Lens' CreateClientVpnRoute Text

The IPv4 address range, in CIDR notation, of the route destination. For example:

To add a route for Internet access, enter 0.0.0.0/0

To add a route for a peered VPC, enter the peered VPC's IPv4 CIDR range

To add a route for an on-premises network, enter the Amazon Web Services Site-to-Site VPN connection's IPv4 CIDR range

To add a route for the local network, enter the client CIDR range

createClientVpnRoute_targetVpcSubnetId :: Lens' CreateClientVpnRoute Text

The ID of the subnet through which you want to route traffic. The specified subnet must be an existing target network of the Client VPN endpoint.

Alternatively, if you're adding a route for the local network, specify local.

Destructuring the Response

data CreateClientVpnRouteResponse

See: newCreateClientVpnRouteResponse smart constructor.

Constructors

CreateClientVpnRouteResponse'

Fields

status :: Maybe ClientVpnRouteStatus

The current state of the route.

httpStatus :: Int

The response's http status code.

newCreateClientVpnRouteResponse

Arguments

:: Int
$sel:httpStatus:CreateClientVpnRouteResponse'

-> CreateClientVpnRouteResponse

Create a value of CreateClientVpnRouteResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:CreateClientVpnRouteResponse', createClientVpnRouteResponse_status - The current state of the route.

$sel:httpStatus:CreateClientVpnRouteResponse', createClientVpnRouteResponse_httpStatus - The response's http status code.

Response Lenses

createClientVpnRouteResponse_status :: Lens' CreateClientVpnRouteResponse (Maybe ClientVpnRouteStatus)

The current state of the route.

createClientVpnRouteResponse_httpStatus :: Lens' CreateClientVpnRouteResponse Int

The response's http status code.