Network.AWS.EC2.CreateCustomerGateway

Description

Provides information to Amazon Web Services about your VPN customer gateway device. The customer gateway is the appliance at your end of the VPN connection. (The device on the Amazon Web Services side of the VPN connection is the virtual private gateway.) You must provide the internet-routable IP address of the customer gateway's external interface. The IP address must be static and can be behind a device performing network address translation (NAT).

For devices that use Border Gateway Protocol (BGP), you can also provide the device's BGP Autonomous System Number (ASN). You can use an existing ASN assigned to your network. If you don't have an ASN already, you can use a private ASN (in the 64512 - 65534 range).

Amazon EC2 supports all 4-byte ASN numbers in the range of 1 - 2147483647, with the exception of the following:

7224 - reserved in the us-east-1 Region

9059 - reserved in the eu-west-1 Region

17943 - reserved in the ap-southeast-1 Region

10124 - reserved in the ap-northeast-1 Region

For more information, see Amazon Web Services Site-to-Site VPN in the Amazon Web Services Site-to-Site VPN User Guide.

To create more than one customer gateway with the same VPN type, IP address, and BGP ASN, specify a unique device name for each customer gateway. Identical requests return information about the existing customer gateway and do not create new customer gateways.

Creating a Request

data CreateCustomerGateway

Contains the parameters for CreateCustomerGateway.

See: newCreateCustomerGateway smart constructor.

Constructors

CreateCustomerGateway'

Fields

certificateArn :: Maybe Text

The Amazon Resource Name (ARN) for the customer gateway certificate.

tagSpecifications :: Maybe [TagSpecification]

The tags to apply to the customer gateway.

deviceName :: Maybe Text

A name for the customer gateway device.

Length Constraints: Up to 255 characters.

publicIp :: Maybe Text

The Internet-routable IP address for the customer gateway's outside interface. The address must be static.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

bgpAsn :: Int

For devices that support BGP, the customer gateway's BGP ASN.

Default: 65000

type' :: GatewayType

The type of VPN connection that this customer gateway supports (ipsec.1).

newCreateCustomerGateway

Arguments

:: Int
$sel:bgpAsn:CreateCustomerGateway'

-> GatewayType
$sel:type':CreateCustomerGateway'

-> CreateCustomerGateway

Create a value of CreateCustomerGateway with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:certificateArn:CreateCustomerGateway', createCustomerGateway_certificateArn - The Amazon Resource Name (ARN) for the customer gateway certificate.

$sel:tagSpecifications:CreateCustomerGateway', createCustomerGateway_tagSpecifications - The tags to apply to the customer gateway.

$sel:deviceName:CreateCustomerGateway', createCustomerGateway_deviceName - A name for the customer gateway device.

Length Constraints: Up to 255 characters.

$sel:publicIp:CreateCustomerGateway', createCustomerGateway_publicIp - The Internet-routable IP address for the customer gateway's outside interface. The address must be static.

$sel:dryRun:CreateCustomerGateway', createCustomerGateway_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:bgpAsn:CreateCustomerGateway', createCustomerGateway_bgpAsn - For devices that support BGP, the customer gateway's BGP ASN.

Default: 65000

$sel:type':CreateCustomerGateway', createCustomerGateway_type - The type of VPN connection that this customer gateway supports (ipsec.1).

Request Lenses

createCustomerGateway_certificateArn :: Lens' CreateCustomerGateway (Maybe Text)

The Amazon Resource Name (ARN) for the customer gateway certificate.

createCustomerGateway_tagSpecifications :: Lens' CreateCustomerGateway (Maybe [TagSpecification])

The tags to apply to the customer gateway.

createCustomerGateway_deviceName :: Lens' CreateCustomerGateway (Maybe Text)

A name for the customer gateway device.

Length Constraints: Up to 255 characters.

createCustomerGateway_publicIp :: Lens' CreateCustomerGateway (Maybe Text)

The Internet-routable IP address for the customer gateway's outside interface. The address must be static.

createCustomerGateway_dryRun :: Lens' CreateCustomerGateway (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

createCustomerGateway_bgpAsn :: Lens' CreateCustomerGateway Int

For devices that support BGP, the customer gateway's BGP ASN.

Default: 65000

createCustomerGateway_type :: Lens' CreateCustomerGateway GatewayType

The type of VPN connection that this customer gateway supports (ipsec.1).

Destructuring the Response

data CreateCustomerGatewayResponse

Contains the output of CreateCustomerGateway.

See: newCreateCustomerGatewayResponse smart constructor.

Constructors

CreateCustomerGatewayResponse'

Fields

customerGateway :: Maybe CustomerGateway

Information about the customer gateway.

httpStatus :: Int

The response's http status code.

newCreateCustomerGatewayResponse

Arguments

:: Int
$sel:httpStatus:CreateCustomerGatewayResponse'

-> CreateCustomerGatewayResponse

Create a value of CreateCustomerGatewayResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:customerGateway:CreateCustomerGatewayResponse', createCustomerGatewayResponse_customerGateway - Information about the customer gateway.

$sel:httpStatus:CreateCustomerGatewayResponse', createCustomerGatewayResponse_httpStatus - The response's http status code.

Response Lenses

createCustomerGatewayResponse_customerGateway :: Lens' CreateCustomerGatewayResponse (Maybe CustomerGateway)

Information about the customer gateway.

createCustomerGatewayResponse_httpStatus :: Lens' CreateCustomerGatewayResponse Int

The response's http status code.