Network.AWS.EC2.CreateDhcpOptions

Description

Creates a set of DHCP options for your VPC. After creating the set, you must associate it with the VPC, causing all existing and new instances that you launch in the VPC to use this set of DHCP options. The following are the individual DHCP options you can specify. For more information about the options, see RFC 2132.

domain-name-servers - The IP addresses of up to four domain name servers, or AmazonProvidedDNS. The default DHCP option set specifies AmazonProvidedDNS. If specifying more than one domain name server, specify the IP addresses in a single parameter, separated by commas. To have your instance receive a custom DNS hostname as specified in domain-name, you must set domain-name-servers to a custom DNS server.

domain-name - If you're using AmazonProvidedDNS in us-east-1, specify ec2.internal. If you're using AmazonProvidedDNS in another Region, specify region.compute.internal (for example, ap-northeast-1.compute.internal). Otherwise, specify a domain name (for example, ExampleCompany.com). This value is used to complete unqualified DNS hostnames. Important: Some Linux operating systems accept multiple domain names separated by spaces. However, Windows and other Linux operating systems treat the value as a single domain, which results in unexpected behavior. If your DHCP options set is associated with a VPC that has instances with multiple operating systems, specify only one domain name.

ntp-servers - The IP addresses of up to four Network Time Protocol (NTP) servers.

netbios-name-servers - The IP addresses of up to four NetBIOS name servers.

netbios-node-type - The NetBIOS node type (1, 2, 4, or 8). We recommend that you specify 2 (broadcast and multicast are not currently supported). For more information about these node types, see RFC 2132.

Your VPC automatically starts out with a set of DHCP options that includes only a DNS server that we provide (AmazonProvidedDNS). If you create a set of options, and if your VPC has an internet gateway, make sure to set the domain-name-servers option either to AmazonProvidedDNS or to a domain name server of your choice. For more information, see DHCP options sets in the Amazon Virtual Private Cloud User Guide.

Creating a Request

data CreateDhcpOptions

See: newCreateDhcpOptions smart constructor.

Constructors

CreateDhcpOptions'

Fields

tagSpecifications :: Maybe [TagSpecification]

The tags to assign to the DHCP option.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

dhcpConfigurations :: [NewDhcpConfiguration]

A DHCP configuration option.

newCreateDhcpOptions :: CreateDhcpOptions

Create a value of CreateDhcpOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tagSpecifications:CreateDhcpOptions', createDhcpOptions_tagSpecifications - The tags to assign to the DHCP option.

$sel:dryRun:CreateDhcpOptions', createDhcpOptions_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:dhcpConfigurations:CreateDhcpOptions', createDhcpOptions_dhcpConfigurations - A DHCP configuration option.

Request Lenses

createDhcpOptions_tagSpecifications :: Lens' CreateDhcpOptions (Maybe [TagSpecification])

The tags to assign to the DHCP option.

createDhcpOptions_dryRun :: Lens' CreateDhcpOptions (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

createDhcpOptions_dhcpConfigurations :: Lens' CreateDhcpOptions [NewDhcpConfiguration]

A DHCP configuration option.

Destructuring the Response

data CreateDhcpOptionsResponse

See: newCreateDhcpOptionsResponse smart constructor.

Constructors

CreateDhcpOptionsResponse'

Fields

dhcpOptions :: Maybe DhcpOptions

A set of DHCP options.

httpStatus :: Int

The response's http status code.

newCreateDhcpOptionsResponse

Arguments

:: Int
$sel:httpStatus:CreateDhcpOptionsResponse'

-> CreateDhcpOptionsResponse

Create a value of CreateDhcpOptionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dhcpOptions:CreateDhcpOptionsResponse', createDhcpOptionsResponse_dhcpOptions - A set of DHCP options.

$sel:httpStatus:CreateDhcpOptionsResponse', createDhcpOptionsResponse_httpStatus - The response's http status code.

Response Lenses

createDhcpOptionsResponse_dhcpOptions :: Lens' CreateDhcpOptionsResponse (Maybe DhcpOptions)

A set of DHCP options.

createDhcpOptionsResponse_httpStatus :: Lens' CreateDhcpOptionsResponse Int

The response's http status code.