Network.AWS.EC2.CreateFleet

Description

Launches an EC2 Fleet.

You can create a single EC2 Fleet that includes multiple launch specifications that vary by instance type, AMI, Availability Zone, or subnet.

For more information, see Launching an EC2 Fleet in the Amazon EC2 User Guide.

Creating a Request

data CreateFleet

See: newCreateFleet smart constructor.

Constructors

CreateFleet'

Fields

context :: Maybe Text

Reserved.

clientToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see Ensuring Idempotency.

spotOptions :: Maybe SpotOptionsRequest

Describes the configuration of Spot Instances in an EC2 Fleet.

excessCapacityTerminationPolicy :: Maybe FleetExcessCapacityTerminationPolicy

Indicates whether running instances should be terminated if the total target capacity of the EC2 Fleet is decreased below the current size of the EC2 Fleet.

onDemandOptions :: Maybe OnDemandOptionsRequest

Describes the configuration of On-Demand Instances in an EC2 Fleet.

tagSpecifications :: Maybe [TagSpecification]

The key-value pair for tagging the EC2 Fleet request on creation. For more information, see Tagging your resources.

If the fleet type is instant, specify a resource type of fleet to tag the fleet or instance to tag the instances at launch.

If the fleet type is maintain or request, specify a resource type of fleet to tag the fleet. You cannot specify a resource type of instance. To tag instances at launch, specify the tags in a launch template.

validUntil :: Maybe ISO8601

The end date and time of the request, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ). At this point, no new EC2 Fleet requests are placed or able to fulfill the request. If no value is specified, the request remains until you cancel it.

terminateInstancesWithExpiration :: Maybe Bool

Indicates whether running instances should be terminated when the EC2 Fleet expires.

type' :: Maybe FleetType

The fleet type. The default value is maintain.

maintain - The EC2 Fleet places an asynchronous request for your desired capacity, and continues to maintain your desired Spot capacity by replenishing interrupted Spot Instances.

request - The EC2 Fleet places an asynchronous one-time request for your desired capacity, but does submit Spot requests in alternative capacity pools if Spot capacity is unavailable, and does not maintain Spot capacity if Spot Instances are interrupted.

instant - The EC2 Fleet places a synchronous one-time request for your desired capacity, and returns errors for any instances that could not be launched.

For more information, see EC2 Fleet request types in the Amazon EC2 User Guide.

validFrom :: Maybe ISO8601

The start date and time of the request, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ). The default is to start fulfilling the request immediately.

replaceUnhealthyInstances :: Maybe Bool

Indicates whether EC2 Fleet should replace unhealthy Spot Instances. Supported only for fleets of type maintain. For more information, see EC2 Fleet health checks in the Amazon EC2 User Guide.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

launchTemplateConfigs :: [FleetLaunchTemplateConfigRequest]

The configuration for the EC2 Fleet.

targetCapacitySpecification :: TargetCapacitySpecificationRequest

The number of units to request.

newCreateFleet

Arguments

:: TargetCapacitySpecificationRequest
$sel:targetCapacitySpecification:CreateFleet'

-> CreateFleet

Create a value of CreateFleet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:context:CreateFleet', createFleet_context - Reserved.

$sel:clientToken:CreateFleet', createFleet_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see Ensuring Idempotency.

$sel:spotOptions:CreateFleet', createFleet_spotOptions - Describes the configuration of Spot Instances in an EC2 Fleet.

$sel:excessCapacityTerminationPolicy:CreateFleet', createFleet_excessCapacityTerminationPolicy - Indicates whether running instances should be terminated if the total target capacity of the EC2 Fleet is decreased below the current size of the EC2 Fleet.

$sel:onDemandOptions:CreateFleet', createFleet_onDemandOptions - Describes the configuration of On-Demand Instances in an EC2 Fleet.

$sel:tagSpecifications:CreateFleet', createFleet_tagSpecifications - The key-value pair for tagging the EC2 Fleet request on creation. For more information, see Tagging your resources.

If the fleet type is instant, specify a resource type of fleet to tag the fleet or instance to tag the instances at launch.

If the fleet type is maintain or request, specify a resource type of fleet to tag the fleet. You cannot specify a resource type of instance. To tag instances at launch, specify the tags in a launch template.

$sel:validUntil:CreateFleet', createFleet_validUntil - The end date and time of the request, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ). At this point, no new EC2 Fleet requests are placed or able to fulfill the request. If no value is specified, the request remains until you cancel it.

$sel:terminateInstancesWithExpiration:CreateFleet', createFleet_terminateInstancesWithExpiration - Indicates whether running instances should be terminated when the EC2 Fleet expires.

$sel:type':CreateFleet', createFleet_type - The fleet type. The default value is maintain.

maintain - The EC2 Fleet places an asynchronous request for your desired capacity, and continues to maintain your desired Spot capacity by replenishing interrupted Spot Instances.

request - The EC2 Fleet places an asynchronous one-time request for your desired capacity, but does submit Spot requests in alternative capacity pools if Spot capacity is unavailable, and does not maintain Spot capacity if Spot Instances are interrupted.

instant - The EC2 Fleet places a synchronous one-time request for your desired capacity, and returns errors for any instances that could not be launched.

For more information, see EC2 Fleet request types in the Amazon EC2 User Guide.

$sel:validFrom:CreateFleet', createFleet_validFrom - The start date and time of the request, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ). The default is to start fulfilling the request immediately.

$sel:replaceUnhealthyInstances:CreateFleet', createFleet_replaceUnhealthyInstances - Indicates whether EC2 Fleet should replace unhealthy Spot Instances. Supported only for fleets of type maintain. For more information, see EC2 Fleet health checks in the Amazon EC2 User Guide.

$sel:dryRun:CreateFleet', createFleet_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:launchTemplateConfigs:CreateFleet', createFleet_launchTemplateConfigs - The configuration for the EC2 Fleet.

$sel:targetCapacitySpecification:CreateFleet', createFleet_targetCapacitySpecification - The number of units to request.

Request Lenses

createFleet_context :: Lens' CreateFleet (Maybe Text)

Reserved.

createFleet_clientToken :: Lens' CreateFleet (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see Ensuring Idempotency.

createFleet_spotOptions :: Lens' CreateFleet (Maybe SpotOptionsRequest)

Describes the configuration of Spot Instances in an EC2 Fleet.

createFleet_excessCapacityTerminationPolicy :: Lens' CreateFleet (Maybe FleetExcessCapacityTerminationPolicy)

Indicates whether running instances should be terminated if the total target capacity of the EC2 Fleet is decreased below the current size of the EC2 Fleet.

createFleet_onDemandOptions :: Lens' CreateFleet (Maybe OnDemandOptionsRequest)

Describes the configuration of On-Demand Instances in an EC2 Fleet.

createFleet_tagSpecifications :: Lens' CreateFleet (Maybe [TagSpecification])

The key-value pair for tagging the EC2 Fleet request on creation. For more information, see Tagging your resources.

If the fleet type is instant, specify a resource type of fleet to tag the fleet or instance to tag the instances at launch.

If the fleet type is maintain or request, specify a resource type of fleet to tag the fleet. You cannot specify a resource type of instance. To tag instances at launch, specify the tags in a launch template.

createFleet_validUntil :: Lens' CreateFleet (Maybe UTCTime)

The end date and time of the request, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ). At this point, no new EC2 Fleet requests are placed or able to fulfill the request. If no value is specified, the request remains until you cancel it.

createFleet_terminateInstancesWithExpiration :: Lens' CreateFleet (Maybe Bool)

Indicates whether running instances should be terminated when the EC2 Fleet expires.

createFleet_type :: Lens' CreateFleet (Maybe FleetType)

The fleet type. The default value is maintain.

maintain - The EC2 Fleet places an asynchronous request for your desired capacity, and continues to maintain your desired Spot capacity by replenishing interrupted Spot Instances.

request - The EC2 Fleet places an asynchronous one-time request for your desired capacity, but does submit Spot requests in alternative capacity pools if Spot capacity is unavailable, and does not maintain Spot capacity if Spot Instances are interrupted.

instant - The EC2 Fleet places a synchronous one-time request for your desired capacity, and returns errors for any instances that could not be launched.

For more information, see EC2 Fleet request types in the Amazon EC2 User Guide.

createFleet_validFrom :: Lens' CreateFleet (Maybe UTCTime)

The start date and time of the request, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ). The default is to start fulfilling the request immediately.

createFleet_replaceUnhealthyInstances :: Lens' CreateFleet (Maybe Bool)

Indicates whether EC2 Fleet should replace unhealthy Spot Instances. Supported only for fleets of type maintain. For more information, see EC2 Fleet health checks in the Amazon EC2 User Guide.

createFleet_dryRun :: Lens' CreateFleet (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

createFleet_launchTemplateConfigs :: Lens' CreateFleet [FleetLaunchTemplateConfigRequest]

The configuration for the EC2 Fleet.

createFleet_targetCapacitySpecification :: Lens' CreateFleet TargetCapacitySpecificationRequest

The number of units to request.

Destructuring the Response

data CreateFleetResponse

See: newCreateFleetResponse smart constructor.

Constructors

CreateFleetResponse'

Fields

instances :: Maybe [CreateFleetInstance]

Information about the instances that were launched by the fleet. Supported only for fleets of type instant.

fleetId :: Maybe Text

The ID of the EC2 Fleet.

errors :: Maybe [CreateFleetError]

Information about the instances that could not be launched by the fleet. Supported only for fleets of type instant.

httpStatus :: Int

The response's http status code.

newCreateFleetResponse

Arguments

:: Int
$sel:httpStatus:CreateFleetResponse'

-> CreateFleetResponse

Create a value of CreateFleetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instances:CreateFleetResponse', createFleetResponse_instances - Information about the instances that were launched by the fleet. Supported only for fleets of type instant.

$sel:fleetId:CreateFleetResponse', createFleetResponse_fleetId - The ID of the EC2 Fleet.

$sel:errors:CreateFleetResponse', createFleetResponse_errors - Information about the instances that could not be launched by the fleet. Supported only for fleets of type instant.

$sel:httpStatus:CreateFleetResponse', createFleetResponse_httpStatus - The response's http status code.

Response Lenses

createFleetResponse_instances :: Lens' CreateFleetResponse (Maybe [CreateFleetInstance])

Information about the instances that were launched by the fleet. Supported only for fleets of type instant.

createFleetResponse_fleetId :: Lens' CreateFleetResponse (Maybe Text)

The ID of the EC2 Fleet.

createFleetResponse_errors :: Lens' CreateFleetResponse (Maybe [CreateFleetError])

Information about the instances that could not be launched by the fleet. Supported only for fleets of type instant.

createFleetResponse_httpStatus :: Lens' CreateFleetResponse Int

The response's http status code.