Network.AWS.EC2.CreateFlowLogs

Description

Creates one or more flow logs to capture information about IP traffic for a specific network interface, subnet, or VPC.

Flow log data for a monitored network interface is recorded as flow log records, which are log events consisting of fields that describe the traffic flow. For more information, see Flow log records in the Amazon Virtual Private Cloud User Guide.

When publishing to CloudWatch Logs, flow log records are published to a log group, and each network interface has a unique log stream in the log group. When publishing to Amazon S3, flow log records for all of the monitored network interfaces are published to a single log file object that is stored in the specified bucket.

For more information, see VPC Flow Logs in the Amazon Virtual Private Cloud User Guide.

Creating a Request

data CreateFlowLogs

See: newCreateFlowLogs smart constructor.

Constructors

CreateFlowLogs'

Fields

maxAggregationInterval :: Maybe Int

The maximum interval of time during which a flow of packets is captured and aggregated into a flow log record. You can specify 60 seconds (1 minute) or 600 seconds (10 minutes).

When a network interface is attached to a Nitro-based instance, the aggregation interval is always 60 seconds or less, regardless of the value that you specify.

Default: 600

tagSpecifications :: Maybe [TagSpecification]

The tags to apply to the flow logs.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

logDestination :: Maybe Text

Specifies the destination to which the flow log data is to be published. Flow log data can be published to a CloudWatch Logs log group or an Amazon S3 bucket. The value specified for this parameter depends on the value specified for LogDestinationType.

If LogDestinationType is not specified or cloud-watch-logs, specify the Amazon Resource Name (ARN) of the CloudWatch Logs log group. For example, to publish to a log group called my-logs, specify arn:aws:logs:us-east-1:123456789012:log-group:my-logs. Alternatively, use LogGroupName instead.

If LogDestinationType is s3, specify the ARN of the Amazon S3 bucket. You can also specify a subfolder in the bucket. To specify a subfolder in the bucket, use the following ARN format: bucket_ARN/subfolder_name/. For example, to specify a subfolder named my-logs in a bucket named my-bucket, use the following ARN: arn:aws:s3:::my-bucket/my-logs/. You cannot use AWSLogs as a subfolder name. This is a reserved term.

logFormat :: Maybe Text

The fields to include in the flow log record, in the order in which they should appear. For a list of available fields, see Flow log records. If you omit this parameter, the flow log is created using the default format. If you specify this parameter, you must specify at least one field.

Specify the fields using the ${field-id} format, separated by spaces. For the CLI, use single quotation marks (' ') to surround the parameter value.

deliverLogsPermissionArn :: Maybe Text

The ARN for the IAM role that permits Amazon EC2 to publish flow logs to a CloudWatch Logs log group in your account.

If you specify LogDestinationType as s3, do not specify DeliverLogsPermissionArn or LogGroupName.

logGroupName :: Maybe Text

The name of a new or existing CloudWatch Logs log group where Amazon EC2 publishes your flow logs.

If you specify LogDestinationType as s3, do not specify DeliverLogsPermissionArn or LogGroupName.

logDestinationType :: Maybe LogDestinationType

Specifies the type of destination to which the flow log data is to be published. Flow log data can be published to CloudWatch Logs or Amazon S3. To publish flow log data to CloudWatch Logs, specify cloud-watch-logs. To publish flow log data to Amazon S3, specify s3.

If you specify LogDestinationType as s3, do not specify DeliverLogsPermissionArn or LogGroupName.

Default: cloud-watch-logs

clientToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to ensure idempotency.

resourceIds :: [Text]

The ID of the subnet, network interface, or VPC for which you want to create a flow log.

Constraints: Maximum of 1000 resources

resourceType :: FlowLogsResourceType

The type of resource for which to create the flow log. For example, if you specified a VPC ID for the ResourceId property, specify VPC for this property.

trafficType :: TrafficType

The type of traffic to log. You can log traffic that the resource accepts or rejects, or all traffic.

newCreateFlowLogs

Arguments

:: FlowLogsResourceType
$sel:resourceType:CreateFlowLogs'

-> TrafficType
$sel:trafficType:CreateFlowLogs'

-> CreateFlowLogs

Create a value of CreateFlowLogs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxAggregationInterval:CreateFlowLogs', createFlowLogs_maxAggregationInterval - The maximum interval of time during which a flow of packets is captured and aggregated into a flow log record. You can specify 60 seconds (1 minute) or 600 seconds (10 minutes).

When a network interface is attached to a Nitro-based instance, the aggregation interval is always 60 seconds or less, regardless of the value that you specify.

Default: 600

$sel:tagSpecifications:CreateFlowLogs', createFlowLogs_tagSpecifications - The tags to apply to the flow logs.

$sel:dryRun:CreateFlowLogs', createFlowLogs_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:logDestination:CreateFlowLogs', createFlowLogs_logDestination - Specifies the destination to which the flow log data is to be published. Flow log data can be published to a CloudWatch Logs log group or an Amazon S3 bucket. The value specified for this parameter depends on the value specified for LogDestinationType.

If LogDestinationType is not specified or cloud-watch-logs, specify the Amazon Resource Name (ARN) of the CloudWatch Logs log group. For example, to publish to a log group called my-logs, specify arn:aws:logs:us-east-1:123456789012:log-group:my-logs. Alternatively, use LogGroupName instead.

If LogDestinationType is s3, specify the ARN of the Amazon S3 bucket. You can also specify a subfolder in the bucket. To specify a subfolder in the bucket, use the following ARN format: bucket_ARN/subfolder_name/. For example, to specify a subfolder named my-logs in a bucket named my-bucket, use the following ARN: arn:aws:s3:::my-bucket/my-logs/. You cannot use AWSLogs as a subfolder name. This is a reserved term.

$sel:logFormat:CreateFlowLogs', createFlowLogs_logFormat - The fields to include in the flow log record, in the order in which they should appear. For a list of available fields, see Flow log records. If you omit this parameter, the flow log is created using the default format. If you specify this parameter, you must specify at least one field.

Specify the fields using the ${field-id} format, separated by spaces. For the CLI, use single quotation marks (' ') to surround the parameter value.

$sel:deliverLogsPermissionArn:CreateFlowLogs', createFlowLogs_deliverLogsPermissionArn - The ARN for the IAM role that permits Amazon EC2 to publish flow logs to a CloudWatch Logs log group in your account.

If you specify LogDestinationType as s3, do not specify DeliverLogsPermissionArn or LogGroupName.

$sel:logGroupName:CreateFlowLogs', createFlowLogs_logGroupName - The name of a new or existing CloudWatch Logs log group where Amazon EC2 publishes your flow logs.

If you specify LogDestinationType as s3, do not specify DeliverLogsPermissionArn or LogGroupName.

$sel:logDestinationType:CreateFlowLogs', createFlowLogs_logDestinationType - Specifies the type of destination to which the flow log data is to be published. Flow log data can be published to CloudWatch Logs or Amazon S3. To publish flow log data to CloudWatch Logs, specify cloud-watch-logs. To publish flow log data to Amazon S3, specify s3.

If you specify LogDestinationType as s3, do not specify DeliverLogsPermissionArn or LogGroupName.

Default: cloud-watch-logs

$sel:clientToken:CreateFlowLogs', createFlowLogs_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to ensure idempotency.

$sel:resourceIds:CreateFlowLogs', createFlowLogs_resourceIds - The ID of the subnet, network interface, or VPC for which you want to create a flow log.

Constraints: Maximum of 1000 resources

$sel:resourceType:CreateFlowLogs', createFlowLogs_resourceType - The type of resource for which to create the flow log. For example, if you specified a VPC ID for the ResourceId property, specify VPC for this property.

$sel:trafficType:CreateFlowLogs', createFlowLogs_trafficType - The type of traffic to log. You can log traffic that the resource accepts or rejects, or all traffic.

Request Lenses

createFlowLogs_maxAggregationInterval :: Lens' CreateFlowLogs (Maybe Int)

The maximum interval of time during which a flow of packets is captured and aggregated into a flow log record. You can specify 60 seconds (1 minute) or 600 seconds (10 minutes).

When a network interface is attached to a Nitro-based instance, the aggregation interval is always 60 seconds or less, regardless of the value that you specify.

Default: 600

createFlowLogs_tagSpecifications :: Lens' CreateFlowLogs (Maybe [TagSpecification])

The tags to apply to the flow logs.

createFlowLogs_dryRun :: Lens' CreateFlowLogs (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

createFlowLogs_logDestination :: Lens' CreateFlowLogs (Maybe Text)

Specifies the destination to which the flow log data is to be published. Flow log data can be published to a CloudWatch Logs log group or an Amazon S3 bucket. The value specified for this parameter depends on the value specified for LogDestinationType.

If LogDestinationType is not specified or cloud-watch-logs, specify the Amazon Resource Name (ARN) of the CloudWatch Logs log group. For example, to publish to a log group called my-logs, specify arn:aws:logs:us-east-1:123456789012:log-group:my-logs. Alternatively, use LogGroupName instead.

If LogDestinationType is s3, specify the ARN of the Amazon S3 bucket. You can also specify a subfolder in the bucket. To specify a subfolder in the bucket, use the following ARN format: bucket_ARN/subfolder_name/. For example, to specify a subfolder named my-logs in a bucket named my-bucket, use the following ARN: arn:aws:s3:::my-bucket/my-logs/. You cannot use AWSLogs as a subfolder name. This is a reserved term.

createFlowLogs_logFormat :: Lens' CreateFlowLogs (Maybe Text)

The fields to include in the flow log record, in the order in which they should appear. For a list of available fields, see Flow log records. If you omit this parameter, the flow log is created using the default format. If you specify this parameter, you must specify at least one field.

Specify the fields using the ${field-id} format, separated by spaces. For the CLI, use single quotation marks (' ') to surround the parameter value.

createFlowLogs_deliverLogsPermissionArn :: Lens' CreateFlowLogs (Maybe Text)

The ARN for the IAM role that permits Amazon EC2 to publish flow logs to a CloudWatch Logs log group in your account.

If you specify LogDestinationType as s3, do not specify DeliverLogsPermissionArn or LogGroupName.

createFlowLogs_logGroupName :: Lens' CreateFlowLogs (Maybe Text)

The name of a new or existing CloudWatch Logs log group where Amazon EC2 publishes your flow logs.

If you specify LogDestinationType as s3, do not specify DeliverLogsPermissionArn or LogGroupName.

createFlowLogs_logDestinationType :: Lens' CreateFlowLogs (Maybe LogDestinationType)

Specifies the type of destination to which the flow log data is to be published. Flow log data can be published to CloudWatch Logs or Amazon S3. To publish flow log data to CloudWatch Logs, specify cloud-watch-logs. To publish flow log data to Amazon S3, specify s3.

If you specify LogDestinationType as s3, do not specify DeliverLogsPermissionArn or LogGroupName.

Default: cloud-watch-logs

createFlowLogs_clientToken :: Lens' CreateFlowLogs (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to ensure idempotency.

createFlowLogs_resourceIds :: Lens' CreateFlowLogs [Text]

The ID of the subnet, network interface, or VPC for which you want to create a flow log.

Constraints: Maximum of 1000 resources

createFlowLogs_resourceType :: Lens' CreateFlowLogs FlowLogsResourceType

The type of resource for which to create the flow log. For example, if you specified a VPC ID for the ResourceId property, specify VPC for this property.

createFlowLogs_trafficType :: Lens' CreateFlowLogs TrafficType

The type of traffic to log. You can log traffic that the resource accepts or rejects, or all traffic.

Destructuring the Response

data CreateFlowLogsResponse

See: newCreateFlowLogsResponse smart constructor.

Constructors

CreateFlowLogsResponse'

Fields

unsuccessful :: Maybe [UnsuccessfulItem]

Information about the flow logs that could not be created successfully.

flowLogIds :: Maybe [Text]

The IDs of the flow logs.

clientToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

httpStatus :: Int

The response's http status code.

newCreateFlowLogsResponse

Arguments

:: Int
$sel:httpStatus:CreateFlowLogsResponse'

-> CreateFlowLogsResponse

Create a value of CreateFlowLogsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:unsuccessful:CreateFlowLogsResponse', createFlowLogsResponse_unsuccessful - Information about the flow logs that could not be created successfully.

$sel:flowLogIds:CreateFlowLogsResponse', createFlowLogsResponse_flowLogIds - The IDs of the flow logs.

$sel:clientToken:CreateFlowLogs', createFlowLogsResponse_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

$sel:httpStatus:CreateFlowLogsResponse', createFlowLogsResponse_httpStatus - The response's http status code.

Response Lenses

createFlowLogsResponse_unsuccessful :: Lens' CreateFlowLogsResponse (Maybe [UnsuccessfulItem])

Information about the flow logs that could not be created successfully.

createFlowLogsResponse_flowLogIds :: Lens' CreateFlowLogsResponse (Maybe [Text])

The IDs of the flow logs.

createFlowLogsResponse_clientToken :: Lens' CreateFlowLogsResponse (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

createFlowLogsResponse_httpStatus :: Lens' CreateFlowLogsResponse Int

The response's http status code.