Network.AWS.EC2.CreateFpgaImage

Description

Creates an Amazon FPGA Image (AFI) from the specified design checkpoint (DCP).

The create operation is asynchronous. To verify that the AFI is ready for use, check the output logs.

An AFI contains the FPGA bitstream that is ready to download to an FPGA. You can securely deploy an AFI on multiple FPGA-accelerated instances. For more information, see the AWS FPGA Hardware Development Kit.

Creating a Request

data CreateFpgaImage

See: newCreateFpgaImage smart constructor.

Constructors

CreateFpgaImage'

Fields

tagSpecifications :: Maybe [TagSpecification]

The tags to apply to the FPGA image during creation.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

name :: Maybe Text

A name for the AFI.

description :: Maybe Text

A description for the AFI.

logsStorageLocation :: Maybe StorageLocation

The location in Amazon S3 for the output logs.

clientToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see Ensuring Idempotency.

inputStorageLocation :: StorageLocation

The location of the encrypted design checkpoint in Amazon S3. The input must be a tarball.

newCreateFpgaImage

Arguments

:: StorageLocation
$sel:inputStorageLocation:CreateFpgaImage'

-> CreateFpgaImage

Create a value of CreateFpgaImage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tagSpecifications:CreateFpgaImage', createFpgaImage_tagSpecifications - The tags to apply to the FPGA image during creation.

$sel:dryRun:CreateFpgaImage', createFpgaImage_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:name:CreateFpgaImage', createFpgaImage_name - A name for the AFI.

$sel:description:CreateFpgaImage', createFpgaImage_description - A description for the AFI.

$sel:logsStorageLocation:CreateFpgaImage', createFpgaImage_logsStorageLocation - The location in Amazon S3 for the output logs.

$sel:clientToken:CreateFpgaImage', createFpgaImage_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see Ensuring Idempotency.

$sel:inputStorageLocation:CreateFpgaImage', createFpgaImage_inputStorageLocation - The location of the encrypted design checkpoint in Amazon S3. The input must be a tarball.

Request Lenses

createFpgaImage_tagSpecifications :: Lens' CreateFpgaImage (Maybe [TagSpecification])

The tags to apply to the FPGA image during creation.

createFpgaImage_dryRun :: Lens' CreateFpgaImage (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

createFpgaImage_name :: Lens' CreateFpgaImage (Maybe Text)

A name for the AFI.

createFpgaImage_description :: Lens' CreateFpgaImage (Maybe Text)

A description for the AFI.

createFpgaImage_logsStorageLocation :: Lens' CreateFpgaImage (Maybe StorageLocation)

The location in Amazon S3 for the output logs.

createFpgaImage_clientToken :: Lens' CreateFpgaImage (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see Ensuring Idempotency.

createFpgaImage_inputStorageLocation :: Lens' CreateFpgaImage StorageLocation

The location of the encrypted design checkpoint in Amazon S3. The input must be a tarball.

Destructuring the Response

data CreateFpgaImageResponse

See: newCreateFpgaImageResponse smart constructor.

Constructors

CreateFpgaImageResponse'

Fields

fpgaImageGlobalId :: Maybe Text

The global FPGA image identifier (AGFI ID).

fpgaImageId :: Maybe Text

The FPGA image identifier (AFI ID).

httpStatus :: Int

The response's http status code.

newCreateFpgaImageResponse

Arguments

:: Int
$sel:httpStatus:CreateFpgaImageResponse'

-> CreateFpgaImageResponse

Create a value of CreateFpgaImageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fpgaImageGlobalId:CreateFpgaImageResponse', createFpgaImageResponse_fpgaImageGlobalId - The global FPGA image identifier (AGFI ID).

$sel:fpgaImageId:CreateFpgaImageResponse', createFpgaImageResponse_fpgaImageId - The FPGA image identifier (AFI ID).

$sel:httpStatus:CreateFpgaImageResponse', createFpgaImageResponse_httpStatus - The response's http status code.

Response Lenses

createFpgaImageResponse_fpgaImageGlobalId :: Lens' CreateFpgaImageResponse (Maybe Text)

The global FPGA image identifier (AGFI ID).

createFpgaImageResponse_fpgaImageId :: Lens' CreateFpgaImageResponse (Maybe Text)

The FPGA image identifier (AFI ID).

createFpgaImageResponse_httpStatus :: Lens' CreateFpgaImageResponse Int

The response's http status code.