Network.AWS.EC2.CreateInstanceEventWindow

Description

Creates an event window in which scheduled events for the associated Amazon EC2 instances can run.

You can define either a set of time ranges or a cron expression when creating the event window, but not both. All event window times are in UTC.

You can create up to 200 event windows per Amazon Web Services Region.

When you create the event window, targets (instance IDs, Dedicated Host IDs, or tags) are not yet associated with it. To ensure that the event window can be used, you must associate one or more targets with it by using the AssociateInstanceEventWindow API.

Event windows are applicable only for scheduled events that stop, reboot, or terminate instances.

Event windows are not applicable for:

Expedited scheduled events and network maintenance events.

Unscheduled maintenance such as AutoRecovery and unplanned reboots.

For more information, see Define event windows for scheduled events in the Amazon EC2 User Guide.

Creating a Request

data CreateInstanceEventWindow

See: newCreateInstanceEventWindow smart constructor.

Constructors

CreateInstanceEventWindow'

Fields

tagSpecifications :: Maybe [TagSpecification]

The tags to apply to the event window.

name :: Maybe Text

The name of the event window.

cronExpression :: Maybe Text

The cron expression for the event window, for example, * 0-4,20-23 * * 1,5. If you specify a cron expression, you can't specify a time range.

Constraints:

Only hour and day of the week values are supported.

For day of the week values, you can specify either integers 0 through 6, or alternative single values SUN through SAT.

The minute, month, and year must be specified by *.

The hour value must be one or a multiple range, for example, 0-4 or 0-4,20-23.

Each hour range must be >= 2 hours, for example, 0-2 or 20-23.

The event window must be >= 4 hours. The combined total time ranges in the event window must be >= 4 hours.

For more information about cron expressions, see cron on the Wikipedia website.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

timeRanges :: Maybe [InstanceEventWindowTimeRangeRequest]

The time range for the event window. If you specify a time range, you can't specify a cron expression.

newCreateInstanceEventWindow :: CreateInstanceEventWindow

Create a value of CreateInstanceEventWindow with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tagSpecifications:CreateInstanceEventWindow', createInstanceEventWindow_tagSpecifications - The tags to apply to the event window.

$sel:name:CreateInstanceEventWindow', createInstanceEventWindow_name - The name of the event window.

$sel:cronExpression:CreateInstanceEventWindow', createInstanceEventWindow_cronExpression - The cron expression for the event window, for example, * 0-4,20-23 * * 1,5. If you specify a cron expression, you can't specify a time range.

Constraints:

Only hour and day of the week values are supported.

For day of the week values, you can specify either integers 0 through 6, or alternative single values SUN through SAT.

The minute, month, and year must be specified by *.

The hour value must be one or a multiple range, for example, 0-4 or 0-4,20-23.

Each hour range must be >= 2 hours, for example, 0-2 or 20-23.

The event window must be >= 4 hours. The combined total time ranges in the event window must be >= 4 hours.

For more information about cron expressions, see cron on the Wikipedia website.

$sel:dryRun:CreateInstanceEventWindow', createInstanceEventWindow_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:timeRanges:CreateInstanceEventWindow', createInstanceEventWindow_timeRanges - The time range for the event window. If you specify a time range, you can't specify a cron expression.

Request Lenses

createInstanceEventWindow_tagSpecifications :: Lens' CreateInstanceEventWindow (Maybe [TagSpecification])

The tags to apply to the event window.

createInstanceEventWindow_name :: Lens' CreateInstanceEventWindow (Maybe Text)

The name of the event window.

createInstanceEventWindow_cronExpression :: Lens' CreateInstanceEventWindow (Maybe Text)

The cron expression for the event window, for example, * 0-4,20-23 * * 1,5. If you specify a cron expression, you can't specify a time range.

Constraints:

Only hour and day of the week values are supported.

For day of the week values, you can specify either integers 0 through 6, or alternative single values SUN through SAT.

The minute, month, and year must be specified by *.

The hour value must be one or a multiple range, for example, 0-4 or 0-4,20-23.

Each hour range must be >= 2 hours, for example, 0-2 or 20-23.

The event window must be >= 4 hours. The combined total time ranges in the event window must be >= 4 hours.

For more information about cron expressions, see cron on the Wikipedia website.

createInstanceEventWindow_dryRun :: Lens' CreateInstanceEventWindow (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

createInstanceEventWindow_timeRanges :: Lens' CreateInstanceEventWindow (Maybe [InstanceEventWindowTimeRangeRequest])

The time range for the event window. If you specify a time range, you can't specify a cron expression.

Destructuring the Response

data CreateInstanceEventWindowResponse

See: newCreateInstanceEventWindowResponse smart constructor.

Constructors

CreateInstanceEventWindowResponse'

Fields

instanceEventWindow :: Maybe InstanceEventWindow

Information about the event window.

httpStatus :: Int

The response's http status code.

newCreateInstanceEventWindowResponse

Arguments

:: Int
$sel:httpStatus:CreateInstanceEventWindowResponse'

-> CreateInstanceEventWindowResponse

Create a value of CreateInstanceEventWindowResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceEventWindow:CreateInstanceEventWindowResponse', createInstanceEventWindowResponse_instanceEventWindow - Information about the event window.

$sel:httpStatus:CreateInstanceEventWindowResponse', createInstanceEventWindowResponse_httpStatus - The response's http status code.

Response Lenses

createInstanceEventWindowResponse_instanceEventWindow :: Lens' CreateInstanceEventWindowResponse (Maybe InstanceEventWindow)

Information about the event window.

createInstanceEventWindowResponse_httpStatus :: Lens' CreateInstanceEventWindowResponse Int

The response's http status code.