Network.AWS.EC2.CreateInstanceExportTask

Description

Exports a running or stopped instance to an Amazon S3 bucket.

For information about the supported operating systems, image formats, and known limitations for the types of instances you can export, see Exporting an instance as a VM Using VM Import/Export in the VM Import/Export User Guide.

Creating a Request

data CreateInstanceExportTask

See: newCreateInstanceExportTask smart constructor.

Constructors

CreateInstanceExportTask'

Fields

tagSpecifications :: Maybe [TagSpecification]

The tags to apply to the export instance task during creation.

description :: Maybe Text

A description for the conversion task or the resource being exported. The maximum length is 255 characters.

exportToS3Task :: ExportToS3TaskSpecification

The format and location for an export instance task.

instanceId :: Text

The ID of the instance.

targetEnvironment :: ExportEnvironment

The target virtualization environment.

newCreateInstanceExportTask

Arguments

:: ExportToS3TaskSpecification
$sel:exportToS3Task:CreateInstanceExportTask'

-> Text
$sel:instanceId:CreateInstanceExportTask'

-> ExportEnvironment
$sel:targetEnvironment:CreateInstanceExportTask'

-> CreateInstanceExportTask

Create a value of CreateInstanceExportTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tagSpecifications:CreateInstanceExportTask', createInstanceExportTask_tagSpecifications - The tags to apply to the export instance task during creation.

$sel:description:CreateInstanceExportTask', createInstanceExportTask_description - A description for the conversion task or the resource being exported. The maximum length is 255 characters.

$sel:exportToS3Task:CreateInstanceExportTask', createInstanceExportTask_exportToS3Task - The format and location for an export instance task.

$sel:instanceId:CreateInstanceExportTask', createInstanceExportTask_instanceId - The ID of the instance.

$sel:targetEnvironment:CreateInstanceExportTask', createInstanceExportTask_targetEnvironment - The target virtualization environment.

Request Lenses

createInstanceExportTask_tagSpecifications :: Lens' CreateInstanceExportTask (Maybe [TagSpecification])

The tags to apply to the export instance task during creation.

createInstanceExportTask_description :: Lens' CreateInstanceExportTask (Maybe Text)

A description for the conversion task or the resource being exported. The maximum length is 255 characters.

createInstanceExportTask_exportToS3Task :: Lens' CreateInstanceExportTask ExportToS3TaskSpecification

The format and location for an export instance task.

createInstanceExportTask_instanceId :: Lens' CreateInstanceExportTask Text

The ID of the instance.

createInstanceExportTask_targetEnvironment :: Lens' CreateInstanceExportTask ExportEnvironment

The target virtualization environment.

Destructuring the Response

data CreateInstanceExportTaskResponse

See: newCreateInstanceExportTaskResponse smart constructor.

Constructors

CreateInstanceExportTaskResponse'

Fields

exportTask :: Maybe ExportTask

Information about the export instance task.

httpStatus :: Int

The response's http status code.

newCreateInstanceExportTaskResponse

Arguments

:: Int
$sel:httpStatus:CreateInstanceExportTaskResponse'

-> CreateInstanceExportTaskResponse

Create a value of CreateInstanceExportTaskResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:exportTask:CreateInstanceExportTaskResponse', createInstanceExportTaskResponse_exportTask - Information about the export instance task.

$sel:httpStatus:CreateInstanceExportTaskResponse', createInstanceExportTaskResponse_httpStatus - The response's http status code.

Response Lenses

createInstanceExportTaskResponse_exportTask :: Lens' CreateInstanceExportTaskResponse (Maybe ExportTask)

Information about the export instance task.

createInstanceExportTaskResponse_httpStatus :: Lens' CreateInstanceExportTaskResponse Int

The response's http status code.