Network.AWS.EC2.CreateLaunchTemplate

Description

Creates a launch template. A launch template contains the parameters to launch an instance. When you launch an instance using RunInstances, you can specify a launch template instead of providing the launch parameters in the request. For more information, see Launching an instance from a launch template in the Amazon Elastic Compute Cloud User Guide.

Creating a Request

data CreateLaunchTemplate

See: newCreateLaunchTemplate smart constructor.

Constructors

CreateLaunchTemplate'

Fields

clientToken :: Maybe Text

Unique, case-sensitive identifier you provide to ensure the idempotency of the request. For more information, see Ensuring Idempotency.

Constraint: Maximum 128 ASCII characters.

versionDescription :: Maybe Text

A description for the first version of the launch template.

tagSpecifications :: Maybe [TagSpecification]

The tags to apply to the launch template during creation.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

launchTemplateName :: Text

A name for the launch template.

launchTemplateData :: RequestLaunchTemplateData

The information for the launch template.

newCreateLaunchTemplate

Arguments

:: Text
$sel:launchTemplateName:CreateLaunchTemplate'

-> RequestLaunchTemplateData
$sel:launchTemplateData:CreateLaunchTemplate'

-> CreateLaunchTemplate

Create a value of CreateLaunchTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateLaunchTemplate', createLaunchTemplate_clientToken - Unique, case-sensitive identifier you provide to ensure the idempotency of the request. For more information, see Ensuring Idempotency.

Constraint: Maximum 128 ASCII characters.

$sel:versionDescription:CreateLaunchTemplate', createLaunchTemplate_versionDescription - A description for the first version of the launch template.

$sel:tagSpecifications:CreateLaunchTemplate', createLaunchTemplate_tagSpecifications - The tags to apply to the launch template during creation.

$sel:dryRun:CreateLaunchTemplate', createLaunchTemplate_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:launchTemplateName:CreateLaunchTemplate', createLaunchTemplate_launchTemplateName - A name for the launch template.

$sel:launchTemplateData:CreateLaunchTemplate', createLaunchTemplate_launchTemplateData - The information for the launch template.

Request Lenses

createLaunchTemplate_clientToken :: Lens' CreateLaunchTemplate (Maybe Text)

Unique, case-sensitive identifier you provide to ensure the idempotency of the request. For more information, see Ensuring Idempotency.

Constraint: Maximum 128 ASCII characters.

createLaunchTemplate_versionDescription :: Lens' CreateLaunchTemplate (Maybe Text)

A description for the first version of the launch template.

createLaunchTemplate_tagSpecifications :: Lens' CreateLaunchTemplate (Maybe [TagSpecification])

The tags to apply to the launch template during creation.

createLaunchTemplate_dryRun :: Lens' CreateLaunchTemplate (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

createLaunchTemplate_launchTemplateName :: Lens' CreateLaunchTemplate Text

A name for the launch template.

createLaunchTemplate_launchTemplateData :: Lens' CreateLaunchTemplate RequestLaunchTemplateData

The information for the launch template.

Destructuring the Response

data CreateLaunchTemplateResponse

See: newCreateLaunchTemplateResponse smart constructor.

Constructors

CreateLaunchTemplateResponse'

Fields

warning :: Maybe ValidationWarning

If the launch template contains parameters or parameter combinations that are not valid, an error code and an error message are returned for each issue that's found.

launchTemplate :: Maybe LaunchTemplate

Information about the launch template.

httpStatus :: Int

The response's http status code.

newCreateLaunchTemplateResponse

Arguments

:: Int
$sel:httpStatus:CreateLaunchTemplateResponse'

-> CreateLaunchTemplateResponse

Create a value of CreateLaunchTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:warning:CreateLaunchTemplateResponse', createLaunchTemplateResponse_warning - If the launch template contains parameters or parameter combinations that are not valid, an error code and an error message are returned for each issue that's found.

$sel:launchTemplate:CreateLaunchTemplateResponse', createLaunchTemplateResponse_launchTemplate - Information about the launch template.

$sel:httpStatus:CreateLaunchTemplateResponse', createLaunchTemplateResponse_httpStatus - The response's http status code.

Response Lenses

createLaunchTemplateResponse_warning :: Lens' CreateLaunchTemplateResponse (Maybe ValidationWarning)

If the launch template contains parameters or parameter combinations that are not valid, an error code and an error message are returned for each issue that's found.

createLaunchTemplateResponse_launchTemplate :: Lens' CreateLaunchTemplateResponse (Maybe LaunchTemplate)

Information about the launch template.

createLaunchTemplateResponse_httpStatus :: Lens' CreateLaunchTemplateResponse Int

The response's http status code.