Network.AWS.EC2.CreateNetworkInsightsPath

Description

Creates a path to analyze for reachability.

Reachability Analyzer enables you to analyze and debug network reachability between two resources in your virtual private cloud (VPC). For more information, see What is Reachability Analyzer.

Creating a Request

data CreateNetworkInsightsPath

See: newCreateNetworkInsightsPath smart constructor.

Constructors

CreateNetworkInsightsPath'

Fields

tagSpecifications :: Maybe [TagSpecification]

The tags to add to the path.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

destinationIp :: Maybe Text

The IP address of the Amazon Web Services resource that is the destination of the path.

sourceIp :: Maybe Text

The IP address of the Amazon Web Services resource that is the source of the path.

destinationPort :: Maybe Natural

The destination port.

source :: Text

The Amazon Web Services resource that is the source of the path.

destination :: Text

The Amazon Web Services resource that is the destination of the path.

protocol :: Protocol

The protocol.

clientToken :: Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to ensure idempotency.

newCreateNetworkInsightsPath

Arguments

:: Text
$sel:source:CreateNetworkInsightsPath'

-> Text
$sel:destination:CreateNetworkInsightsPath'

-> Protocol
$sel:protocol:CreateNetworkInsightsPath'

-> Text
$sel:clientToken:CreateNetworkInsightsPath'

-> CreateNetworkInsightsPath

Create a value of CreateNetworkInsightsPath with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tagSpecifications:CreateNetworkInsightsPath', createNetworkInsightsPath_tagSpecifications - The tags to add to the path.

$sel:dryRun:CreateNetworkInsightsPath', createNetworkInsightsPath_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:destinationIp:CreateNetworkInsightsPath', createNetworkInsightsPath_destinationIp - The IP address of the Amazon Web Services resource that is the destination of the path.

$sel:sourceIp:CreateNetworkInsightsPath', createNetworkInsightsPath_sourceIp - The IP address of the Amazon Web Services resource that is the source of the path.

$sel:destinationPort:CreateNetworkInsightsPath', createNetworkInsightsPath_destinationPort - The destination port.

$sel:source:CreateNetworkInsightsPath', createNetworkInsightsPath_source - The Amazon Web Services resource that is the source of the path.

$sel:destination:CreateNetworkInsightsPath', createNetworkInsightsPath_destination - The Amazon Web Services resource that is the destination of the path.

$sel:protocol:CreateNetworkInsightsPath', createNetworkInsightsPath_protocol - The protocol.

$sel:clientToken:CreateNetworkInsightsPath', createNetworkInsightsPath_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to ensure idempotency.

Request Lenses

createNetworkInsightsPath_tagSpecifications :: Lens' CreateNetworkInsightsPath (Maybe [TagSpecification])

The tags to add to the path.

createNetworkInsightsPath_dryRun :: Lens' CreateNetworkInsightsPath (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

createNetworkInsightsPath_destinationIp :: Lens' CreateNetworkInsightsPath (Maybe Text)

The IP address of the Amazon Web Services resource that is the destination of the path.

createNetworkInsightsPath_sourceIp :: Lens' CreateNetworkInsightsPath (Maybe Text)

The IP address of the Amazon Web Services resource that is the source of the path.

createNetworkInsightsPath_destinationPort :: Lens' CreateNetworkInsightsPath (Maybe Natural)

The destination port.

createNetworkInsightsPath_source :: Lens' CreateNetworkInsightsPath Text

The Amazon Web Services resource that is the source of the path.

createNetworkInsightsPath_destination :: Lens' CreateNetworkInsightsPath Text

The Amazon Web Services resource that is the destination of the path.

createNetworkInsightsPath_protocol :: Lens' CreateNetworkInsightsPath Protocol

The protocol.

createNetworkInsightsPath_clientToken :: Lens' CreateNetworkInsightsPath Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to ensure idempotency.

Destructuring the Response

data CreateNetworkInsightsPathResponse

See: newCreateNetworkInsightsPathResponse smart constructor.

Constructors

CreateNetworkInsightsPathResponse'

Fields

networkInsightsPath :: Maybe NetworkInsightsPath

Information about the path.

httpStatus :: Int

The response's http status code.

newCreateNetworkInsightsPathResponse

Arguments

:: Int
$sel:httpStatus:CreateNetworkInsightsPathResponse'

-> CreateNetworkInsightsPathResponse

Create a value of CreateNetworkInsightsPathResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:networkInsightsPath:CreateNetworkInsightsPathResponse', createNetworkInsightsPathResponse_networkInsightsPath - Information about the path.

$sel:httpStatus:CreateNetworkInsightsPathResponse', createNetworkInsightsPathResponse_httpStatus - The response's http status code.

Response Lenses

createNetworkInsightsPathResponse_networkInsightsPath :: Lens' CreateNetworkInsightsPathResponse (Maybe NetworkInsightsPath)

Information about the path.

createNetworkInsightsPathResponse_httpStatus :: Lens' CreateNetworkInsightsPathResponse Int

The response's http status code.