Network.AWS.EC2.CreatePlacementGroup

Description

Creates a placement group in which to launch instances. The strategy of the placement group determines how the instances are organized within the group.

A cluster placement group is a logical grouping of instances within a single Availability Zone that benefit from low network latency, high network throughput. A spread placement group places instances on distinct hardware. A partition placement group places groups of instances in different partitions, where instances in one partition do not share the same hardware with instances in another partition.

For more information, see Placement groups in the Amazon EC2 User Guide.

Creating a Request

data CreatePlacementGroup

See: newCreatePlacementGroup smart constructor.

Constructors

CreatePlacementGroup'

Fields

tagSpecifications :: Maybe [TagSpecification]

The tags to apply to the new placement group.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

strategy :: Maybe PlacementStrategy

The placement strategy.

groupName :: Maybe Text

A name for the placement group. Must be unique within the scope of your account for the Region.

Constraints: Up to 255 ASCII characters

partitionCount :: Maybe Int

The number of partitions. Valid only when Strategy is set to partition.

newCreatePlacementGroup :: CreatePlacementGroup

Create a value of CreatePlacementGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tagSpecifications:CreatePlacementGroup', createPlacementGroup_tagSpecifications - The tags to apply to the new placement group.

$sel:dryRun:CreatePlacementGroup', createPlacementGroup_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:strategy:CreatePlacementGroup', createPlacementGroup_strategy - The placement strategy.

$sel:groupName:CreatePlacementGroup', createPlacementGroup_groupName - A name for the placement group. Must be unique within the scope of your account for the Region.

Constraints: Up to 255 ASCII characters

$sel:partitionCount:CreatePlacementGroup', createPlacementGroup_partitionCount - The number of partitions. Valid only when Strategy is set to partition.

Request Lenses

createPlacementGroup_tagSpecifications :: Lens' CreatePlacementGroup (Maybe [TagSpecification])

The tags to apply to the new placement group.

createPlacementGroup_dryRun :: Lens' CreatePlacementGroup (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

createPlacementGroup_strategy :: Lens' CreatePlacementGroup (Maybe PlacementStrategy)

The placement strategy.

createPlacementGroup_groupName :: Lens' CreatePlacementGroup (Maybe Text)

A name for the placement group. Must be unique within the scope of your account for the Region.

Constraints: Up to 255 ASCII characters

createPlacementGroup_partitionCount :: Lens' CreatePlacementGroup (Maybe Int)

The number of partitions. Valid only when Strategy is set to partition.

Destructuring the Response

data CreatePlacementGroupResponse

See: newCreatePlacementGroupResponse smart constructor.

Constructors

CreatePlacementGroupResponse'

Fields

placementGroup :: Maybe PlacementGroup

httpStatus :: Int

The response's http status code.

newCreatePlacementGroupResponse

Arguments

:: Int
$sel:httpStatus:CreatePlacementGroupResponse'

-> CreatePlacementGroupResponse

Create a value of CreatePlacementGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:placementGroup:CreatePlacementGroupResponse', createPlacementGroupResponse_placementGroup - Undocumented member.

$sel:httpStatus:CreatePlacementGroupResponse', createPlacementGroupResponse_httpStatus - The response's http status code.

Response Lenses

createPlacementGroupResponse_placementGroup :: Lens' CreatePlacementGroupResponse (Maybe PlacementGroup)

Undocumented member.

createPlacementGroupResponse_httpStatus :: Lens' CreatePlacementGroupResponse Int

The response's http status code.