Network.AWS.EC2.CreateReservedInstancesListing

Description

Creates a listing for Amazon EC2 Standard Reserved Instances to be sold in the Reserved Instance Marketplace. You can submit one Standard Reserved Instance listing at a time. To get a list of your Standard Reserved Instances, you can use the DescribeReservedInstances operation.

Only Standard Reserved Instances can be sold in the Reserved Instance Marketplace. Convertible Reserved Instances cannot be sold.

The Reserved Instance Marketplace matches sellers who want to resell Standard Reserved Instance capacity that they no longer need with buyers who want to purchase additional capacity. Reserved Instances bought and sold through the Reserved Instance Marketplace work like any other Reserved Instances.

To sell your Standard Reserved Instances, you must first register as a seller in the Reserved Instance Marketplace. After completing the registration process, you can create a Reserved Instance Marketplace listing of some or all of your Standard Reserved Instances, and specify the upfront price to receive for them. Your Standard Reserved Instance listings then become available for purchase. To view the details of your Standard Reserved Instance listing, you can use the DescribeReservedInstancesListings operation.

For more information, see Reserved Instance Marketplace in the Amazon EC2 User Guide.

Creating a Request

data CreateReservedInstancesListing

Contains the parameters for CreateReservedInstancesListing.

See: newCreateReservedInstancesListing smart constructor.

Constructors

CreateReservedInstancesListing'

Fields

clientToken :: Text

Unique, case-sensitive identifier you provide to ensure idempotency of your listings. This helps avoid duplicate listings. For more information, see Ensuring Idempotency.

instanceCount :: Int

The number of instances that are a part of a Reserved Instance account to be listed in the Reserved Instance Marketplace. This number should be less than or equal to the instance count associated with the Reserved Instance ID specified in this call.

priceSchedules :: [PriceScheduleSpecification]

A list specifying the price of the Standard Reserved Instance for each month remaining in the Reserved Instance term.

reservedInstancesId :: Text

The ID of the active Standard Reserved Instance.

newCreateReservedInstancesListing

Arguments

:: Text
$sel:clientToken:CreateReservedInstancesListing'

-> Int
$sel:instanceCount:CreateReservedInstancesListing'

-> Text
$sel:reservedInstancesId:CreateReservedInstancesListing'

-> CreateReservedInstancesListing

Create a value of CreateReservedInstancesListing with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateReservedInstancesListing', createReservedInstancesListing_clientToken - Unique, case-sensitive identifier you provide to ensure idempotency of your listings. This helps avoid duplicate listings. For more information, see Ensuring Idempotency.

$sel:instanceCount:CreateReservedInstancesListing', createReservedInstancesListing_instanceCount - The number of instances that are a part of a Reserved Instance account to be listed in the Reserved Instance Marketplace. This number should be less than or equal to the instance count associated with the Reserved Instance ID specified in this call.

$sel:priceSchedules:CreateReservedInstancesListing', createReservedInstancesListing_priceSchedules - A list specifying the price of the Standard Reserved Instance for each month remaining in the Reserved Instance term.

$sel:reservedInstancesId:CreateReservedInstancesListing', createReservedInstancesListing_reservedInstancesId - The ID of the active Standard Reserved Instance.

Request Lenses

createReservedInstancesListing_clientToken :: Lens' CreateReservedInstancesListing Text

Unique, case-sensitive identifier you provide to ensure idempotency of your listings. This helps avoid duplicate listings. For more information, see Ensuring Idempotency.

createReservedInstancesListing_instanceCount :: Lens' CreateReservedInstancesListing Int

The number of instances that are a part of a Reserved Instance account to be listed in the Reserved Instance Marketplace. This number should be less than or equal to the instance count associated with the Reserved Instance ID specified in this call.

createReservedInstancesListing_priceSchedules :: Lens' CreateReservedInstancesListing [PriceScheduleSpecification]

A list specifying the price of the Standard Reserved Instance for each month remaining in the Reserved Instance term.

createReservedInstancesListing_reservedInstancesId :: Lens' CreateReservedInstancesListing Text

The ID of the active Standard Reserved Instance.

Destructuring the Response

data CreateReservedInstancesListingResponse

Contains the output of CreateReservedInstancesListing.

See: newCreateReservedInstancesListingResponse smart constructor.

Constructors

CreateReservedInstancesListingResponse'

Fields

reservedInstancesListings :: Maybe [ReservedInstancesListing]

Information about the Standard Reserved Instance listing.

httpStatus :: Int

The response's http status code.

newCreateReservedInstancesListingResponse

Arguments

:: Int
$sel:httpStatus:CreateReservedInstancesListingResponse'

-> CreateReservedInstancesListingResponse

Create a value of CreateReservedInstancesListingResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reservedInstancesListings:CreateReservedInstancesListingResponse', createReservedInstancesListingResponse_reservedInstancesListings - Information about the Standard Reserved Instance listing.

$sel:httpStatus:CreateReservedInstancesListingResponse', createReservedInstancesListingResponse_httpStatus - The response's http status code.

Response Lenses

createReservedInstancesListingResponse_reservedInstancesListings :: Lens' CreateReservedInstancesListingResponse (Maybe [ReservedInstancesListing])

Information about the Standard Reserved Instance listing.

createReservedInstancesListingResponse_httpStatus :: Lens' CreateReservedInstancesListingResponse Int

The response's http status code.