Network.AWS.EC2.CreateRoute

Description

Creates a route in a route table within a VPC.

You must specify one of the following targets: internet gateway or virtual private gateway, NAT instance, NAT gateway, VPC peering connection, network interface, egress-only internet gateway, or transit gateway.

When determining how to route traffic, we use the route with the most specific match. For example, traffic is destined for the IPv4 address 192.0.2.3, and the route table includes the following two IPv4 routes:

192.0.2.0/24 (goes to some target A)

192.0.2.0/28 (goes to some target B)

Both routes apply to the traffic destined for 192.0.2.3. However, the second route in the list covers a smaller number of IP addresses and is therefore more specific, so we use that route to determine where to target the traffic.

For more information about route tables, see Route tables in the Amazon Virtual Private Cloud User Guide.

Creating a Request

data CreateRoute

See: newCreateRoute smart constructor.

Constructors

CreateRoute'

Fields

vpcPeeringConnectionId :: Maybe Text

The ID of a VPC peering connection.

instanceId :: Maybe Text

The ID of a NAT instance in your VPC. The operation fails if you specify an instance ID unless exactly one network interface is attached.

egressOnlyInternetGatewayId :: Maybe Text

IPv6 traffic only

The ID of an egress-only internet gateway.

destinationIpv6CidrBlock :: Maybe Text

The IPv6 CIDR block used for the destination match. Routing decisions are based on the most specific match.

localGatewayId :: Maybe Text

The ID of the local gateway.

natGatewayId :: Maybe Text

IPv4 traffic only

The ID of a NAT gateway.

networkInterfaceId :: Maybe Text

The ID of a network interface.

transitGatewayId :: Maybe Text

The ID of a transit gateway.

gatewayId :: Maybe Text

The ID of an internet gateway or virtual private gateway attached to your VPC.

vpcEndpointId :: Maybe Text

The ID of a VPC endpoint. Supported for Gateway Load Balancer endpoints only.

destinationPrefixListId :: Maybe Text

The ID of a prefix list used for the destination match.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

carrierGatewayId :: Maybe Text

The ID of the carrier gateway.

You can only use this option when the VPC contains a subnet which is associated with a Wavelength Zone.

destinationCidrBlock :: Maybe Text

The IPv4 CIDR address block used for the destination match. Routing decisions are based on the most specific match. We modify the specified CIDR block to its canonical form; for example, if you specify 100.68.0.18/18, we modify it to 100.68.0.0/18.

routeTableId :: Text

The ID of the route table for the route.

newCreateRoute

Arguments

:: Text
$sel:routeTableId:CreateRoute'

-> CreateRoute

Create a value of CreateRoute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpcPeeringConnectionId:CreateRoute', createRoute_vpcPeeringConnectionId - The ID of a VPC peering connection.

$sel:instanceId:CreateRoute', createRoute_instanceId - The ID of a NAT instance in your VPC. The operation fails if you specify an instance ID unless exactly one network interface is attached.

$sel:egressOnlyInternetGatewayId:CreateRoute', createRoute_egressOnlyInternetGatewayId - [IPv6 traffic only] The ID of an egress-only internet gateway.

$sel:destinationIpv6CidrBlock:CreateRoute', createRoute_destinationIpv6CidrBlock - The IPv6 CIDR block used for the destination match. Routing decisions are based on the most specific match.

$sel:localGatewayId:CreateRoute', createRoute_localGatewayId - The ID of the local gateway.

$sel:natGatewayId:CreateRoute', createRoute_natGatewayId - [IPv4 traffic only] The ID of a NAT gateway.

$sel:networkInterfaceId:CreateRoute', createRoute_networkInterfaceId - The ID of a network interface.

$sel:transitGatewayId:CreateRoute', createRoute_transitGatewayId - The ID of a transit gateway.

$sel:gatewayId:CreateRoute', createRoute_gatewayId - The ID of an internet gateway or virtual private gateway attached to your VPC.

$sel:vpcEndpointId:CreateRoute', createRoute_vpcEndpointId - The ID of a VPC endpoint. Supported for Gateway Load Balancer endpoints only.

$sel:destinationPrefixListId:CreateRoute', createRoute_destinationPrefixListId - The ID of a prefix list used for the destination match.

$sel:dryRun:CreateRoute', createRoute_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:carrierGatewayId:CreateRoute', createRoute_carrierGatewayId - The ID of the carrier gateway.

You can only use this option when the VPC contains a subnet which is associated with a Wavelength Zone.

$sel:destinationCidrBlock:CreateRoute', createRoute_destinationCidrBlock - The IPv4 CIDR address block used for the destination match. Routing decisions are based on the most specific match. We modify the specified CIDR block to its canonical form; for example, if you specify 100.68.0.18/18, we modify it to 100.68.0.0/18.

$sel:routeTableId:CreateRoute', createRoute_routeTableId - The ID of the route table for the route.

Request Lenses

createRoute_vpcPeeringConnectionId :: Lens' CreateRoute (Maybe Text)

The ID of a VPC peering connection.

createRoute_instanceId :: Lens' CreateRoute (Maybe Text)

The ID of a NAT instance in your VPC. The operation fails if you specify an instance ID unless exactly one network interface is attached.

createRoute_egressOnlyInternetGatewayId :: Lens' CreateRoute (Maybe Text)

IPv6 traffic only

The ID of an egress-only internet gateway.

createRoute_destinationIpv6CidrBlock :: Lens' CreateRoute (Maybe Text)

The IPv6 CIDR block used for the destination match. Routing decisions are based on the most specific match.

createRoute_localGatewayId :: Lens' CreateRoute (Maybe Text)

The ID of the local gateway.

createRoute_natGatewayId :: Lens' CreateRoute (Maybe Text)

IPv4 traffic only

The ID of a NAT gateway.

createRoute_networkInterfaceId :: Lens' CreateRoute (Maybe Text)

The ID of a network interface.

createRoute_transitGatewayId :: Lens' CreateRoute (Maybe Text)

The ID of a transit gateway.

createRoute_gatewayId :: Lens' CreateRoute (Maybe Text)

The ID of an internet gateway or virtual private gateway attached to your VPC.

createRoute_vpcEndpointId :: Lens' CreateRoute (Maybe Text)

The ID of a VPC endpoint. Supported for Gateway Load Balancer endpoints only.

createRoute_destinationPrefixListId :: Lens' CreateRoute (Maybe Text)

The ID of a prefix list used for the destination match.

createRoute_dryRun :: Lens' CreateRoute (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

createRoute_carrierGatewayId :: Lens' CreateRoute (Maybe Text)

The ID of the carrier gateway.

You can only use this option when the VPC contains a subnet which is associated with a Wavelength Zone.

createRoute_destinationCidrBlock :: Lens' CreateRoute (Maybe Text)

The IPv4 CIDR address block used for the destination match. Routing decisions are based on the most specific match. We modify the specified CIDR block to its canonical form; for example, if you specify 100.68.0.18/18, we modify it to 100.68.0.0/18.

createRoute_routeTableId :: Lens' CreateRoute Text

The ID of the route table for the route.

Destructuring the Response

data CreateRouteResponse

See: newCreateRouteResponse smart constructor.

Constructors

CreateRouteResponse'

Fields

return' :: Maybe Bool

Returns true if the request succeeds; otherwise, it returns an error.

httpStatus :: Int

The response's http status code.

newCreateRouteResponse

Arguments

:: Int
$sel:httpStatus:CreateRouteResponse'

-> CreateRouteResponse

Create a value of CreateRouteResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:return':CreateRouteResponse', createRouteResponse_return - Returns true if the request succeeds; otherwise, it returns an error.

$sel:httpStatus:CreateRouteResponse', createRouteResponse_httpStatus - The response's http status code.

Response Lenses

createRouteResponse_return :: Lens' CreateRouteResponse (Maybe Bool)

Returns true if the request succeeds; otherwise, it returns an error.

createRouteResponse_httpStatus :: Lens' CreateRouteResponse Int

The response's http status code.