Network.AWS.EC2.CreateSnapshots

Description

Creates crash-consistent snapshots of multiple EBS volumes and stores the data in S3. Volumes are chosen by specifying an instance. Any attached volumes will produce one snapshot each that is crash-consistent across the instance. Boot volumes can be excluded by changing the parameters.

You can create multi-volume snapshots of instances in a Region and instances on an Outpost. If you create snapshots from an instance in a Region, the snapshots must be stored in the same Region as the instance. If you create snapshots from an instance on an Outpost, the snapshots can be stored on the same Outpost as the instance, or in the Region for that Outpost.

Creating a Request

data CreateSnapshots

See: newCreateSnapshots smart constructor.

Constructors

CreateSnapshots'

Fields

tagSpecifications :: Maybe [TagSpecification]

Tags to apply to every snapshot specified by the instance.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

outpostArn :: Maybe Text

The Amazon Resource Name (ARN) of the Outpost on which to create the local snapshots.

To create snapshots from an instance in a Region, omit this parameter. The snapshots are created in the same Region as the instance.

To create snapshots from an instance on an Outpost and store the snapshots in the Region, omit this parameter. The snapshots are created in the Region for the Outpost.

To create snapshots from an instance on an Outpost and store the snapshots on an Outpost, specify the ARN of the destination Outpost. The snapshots must be created on the same Outpost as the instance.

For more information, see Create multi-volume local snapshots from instances on an Outpost in the Amazon Elastic Compute Cloud User Guide.

copyTagsFromSource :: Maybe CopyTagsFromSource

Copies the tags from the specified volume to corresponding snapshot.

description :: Maybe Text

A description propagated to every snapshot specified by the instance.

instanceSpecification :: InstanceSpecification

The instance to specify which volumes should be included in the snapshots.

newCreateSnapshots

Arguments

:: InstanceSpecification
$sel:instanceSpecification:CreateSnapshots'

-> CreateSnapshots

Create a value of CreateSnapshots with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tagSpecifications:CreateSnapshots', createSnapshots_tagSpecifications - Tags to apply to every snapshot specified by the instance.

$sel:dryRun:CreateSnapshots', createSnapshots_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:outpostArn:CreateSnapshots', createSnapshots_outpostArn - The Amazon Resource Name (ARN) of the Outpost on which to create the local snapshots.

To create snapshots from an instance in a Region, omit this parameter. The snapshots are created in the same Region as the instance.

To create snapshots from an instance on an Outpost and store the snapshots in the Region, omit this parameter. The snapshots are created in the Region for the Outpost.

To create snapshots from an instance on an Outpost and store the snapshots on an Outpost, specify the ARN of the destination Outpost. The snapshots must be created on the same Outpost as the instance.

For more information, see Create multi-volume local snapshots from instances on an Outpost in the Amazon Elastic Compute Cloud User Guide.

$sel:copyTagsFromSource:CreateSnapshots', createSnapshots_copyTagsFromSource - Copies the tags from the specified volume to corresponding snapshot.

$sel:description:CreateSnapshots', createSnapshots_description - A description propagated to every snapshot specified by the instance.

$sel:instanceSpecification:CreateSnapshots', createSnapshots_instanceSpecification - The instance to specify which volumes should be included in the snapshots.

Request Lenses

createSnapshots_tagSpecifications :: Lens' CreateSnapshots (Maybe [TagSpecification])

Tags to apply to every snapshot specified by the instance.

createSnapshots_dryRun :: Lens' CreateSnapshots (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

createSnapshots_outpostArn :: Lens' CreateSnapshots (Maybe Text)

The Amazon Resource Name (ARN) of the Outpost on which to create the local snapshots.

To create snapshots from an instance in a Region, omit this parameter. The snapshots are created in the same Region as the instance.

To create snapshots from an instance on an Outpost and store the snapshots in the Region, omit this parameter. The snapshots are created in the Region for the Outpost.

To create snapshots from an instance on an Outpost and store the snapshots on an Outpost, specify the ARN of the destination Outpost. The snapshots must be created on the same Outpost as the instance.

For more information, see Create multi-volume local snapshots from instances on an Outpost in the Amazon Elastic Compute Cloud User Guide.

createSnapshots_copyTagsFromSource :: Lens' CreateSnapshots (Maybe CopyTagsFromSource)

Copies the tags from the specified volume to corresponding snapshot.

createSnapshots_description :: Lens' CreateSnapshots (Maybe Text)

A description propagated to every snapshot specified by the instance.

createSnapshots_instanceSpecification :: Lens' CreateSnapshots InstanceSpecification

The instance to specify which volumes should be included in the snapshots.

Destructuring the Response

data CreateSnapshotsResponse

See: newCreateSnapshotsResponse smart constructor.

Constructors

CreateSnapshotsResponse'

Fields

snapshots :: Maybe [SnapshotInfo]

List of snapshots.

httpStatus :: Int

The response's http status code.

newCreateSnapshotsResponse

Arguments

:: Int
$sel:httpStatus:CreateSnapshotsResponse'

-> CreateSnapshotsResponse

Create a value of CreateSnapshotsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:snapshots:CreateSnapshotsResponse', createSnapshotsResponse_snapshots - List of snapshots.

$sel:httpStatus:CreateSnapshotsResponse', createSnapshotsResponse_httpStatus - The response's http status code.

Response Lenses

createSnapshotsResponse_snapshots :: Lens' CreateSnapshotsResponse (Maybe [SnapshotInfo])

List of snapshots.

createSnapshotsResponse_httpStatus :: Lens' CreateSnapshotsResponse Int

The response's http status code.