Network.AWS.EC2.CreateSubnet

Description

Creates a subnet in a specified VPC.

You must specify an IPv4 CIDR block for the subnet. After you create a subnet, you can't change its CIDR block. The allowed block size is between a /16 netmask (65,536 IP addresses) and /28 netmask (16 IP addresses). The CIDR block must not overlap with the CIDR block of an existing subnet in the VPC.

If you've associated an IPv6 CIDR block with your VPC, you can create a subnet with an IPv6 CIDR block that uses a /64 prefix length.

Amazon Web Services reserves both the first four and the last IPv4 address in each subnet's CIDR block. They're not available for use.

If you add more than one subnet to a VPC, they're set up in a star topology with a logical router in the middle.

When you stop an instance in a subnet, it retains its private IPv4 address. It's therefore possible to have a subnet with no running instances (they're all stopped), but no remaining IP addresses available.

For more information about subnets, see Your VPC and subnets in the Amazon Virtual Private Cloud User Guide.

Creating a Request

data CreateSubnet

See: newCreateSubnet smart constructor.

Constructors

CreateSubnet'

Fields

ipv6CidrBlock :: Maybe Text

The IPv6 network range for the subnet, in CIDR notation. The subnet size must use a /64 prefix length.

availabilityZoneId :: Maybe Text

The AZ ID or the Local Zone ID of the subnet.

outpostArn :: Maybe Text

The Amazon Resource Name (ARN) of the Outpost. If you specify an Outpost ARN, you must also specify the Availability Zone of the Outpost subnet.

tagSpecifications :: Maybe [TagSpecification]

The tags to assign to the subnet.

availabilityZone :: Maybe Text

The Availability Zone or Local Zone for the subnet.

Default: Amazon Web Services selects one for you. If you create more than one subnet in your VPC, we do not necessarily select a different zone for each subnet.

To create a subnet in a Local Zone, set this value to the Local Zone ID, for example us-west-2-lax-1a. For information about the Regions that support Local Zones, see Available Regions in the Amazon Elastic Compute Cloud User Guide.

To create a subnet in an Outpost, set this value to the Availability Zone for the Outpost and specify the Outpost ARN.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

cidrBlock :: Text

The IPv4 network range for the subnet, in CIDR notation. For example, 10.0.0.0/24. We modify the specified CIDR block to its canonical form; for example, if you specify 100.68.0.18/18, we modify it to 100.68.0.0/18.

vpcId :: Text

The ID of the VPC.

newCreateSubnet

Arguments

:: Text
$sel:cidrBlock:CreateSubnet'

-> Text
$sel:vpcId:CreateSubnet'

-> CreateSubnet

Create a value of CreateSubnet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ipv6CidrBlock:CreateSubnet', createSubnet_ipv6CidrBlock - The IPv6 network range for the subnet, in CIDR notation. The subnet size must use a /64 prefix length.

$sel:availabilityZoneId:CreateSubnet', createSubnet_availabilityZoneId - The AZ ID or the Local Zone ID of the subnet.

$sel:outpostArn:CreateSubnet', createSubnet_outpostArn - The Amazon Resource Name (ARN) of the Outpost. If you specify an Outpost ARN, you must also specify the Availability Zone of the Outpost subnet.

$sel:tagSpecifications:CreateSubnet', createSubnet_tagSpecifications - The tags to assign to the subnet.

$sel:availabilityZone:CreateSubnet', createSubnet_availabilityZone - The Availability Zone or Local Zone for the subnet.

Default: Amazon Web Services selects one for you. If you create more than one subnet in your VPC, we do not necessarily select a different zone for each subnet.

To create a subnet in a Local Zone, set this value to the Local Zone ID, for example us-west-2-lax-1a. For information about the Regions that support Local Zones, see Available Regions in the Amazon Elastic Compute Cloud User Guide.

To create a subnet in an Outpost, set this value to the Availability Zone for the Outpost and specify the Outpost ARN.

$sel:dryRun:CreateSubnet', createSubnet_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:cidrBlock:CreateSubnet', createSubnet_cidrBlock - The IPv4 network range for the subnet, in CIDR notation. For example, 10.0.0.0/24. We modify the specified CIDR block to its canonical form; for example, if you specify 100.68.0.18/18, we modify it to 100.68.0.0/18.

$sel:vpcId:CreateSubnet', createSubnet_vpcId - The ID of the VPC.

Request Lenses

createSubnet_ipv6CidrBlock :: Lens' CreateSubnet (Maybe Text)

The IPv6 network range for the subnet, in CIDR notation. The subnet size must use a /64 prefix length.

createSubnet_availabilityZoneId :: Lens' CreateSubnet (Maybe Text)

The AZ ID or the Local Zone ID of the subnet.

createSubnet_outpostArn :: Lens' CreateSubnet (Maybe Text)

The Amazon Resource Name (ARN) of the Outpost. If you specify an Outpost ARN, you must also specify the Availability Zone of the Outpost subnet.

createSubnet_tagSpecifications :: Lens' CreateSubnet (Maybe [TagSpecification])

The tags to assign to the subnet.

createSubnet_availabilityZone :: Lens' CreateSubnet (Maybe Text)

The Availability Zone or Local Zone for the subnet.

Default: Amazon Web Services selects one for you. If you create more than one subnet in your VPC, we do not necessarily select a different zone for each subnet.

To create a subnet in a Local Zone, set this value to the Local Zone ID, for example us-west-2-lax-1a. For information about the Regions that support Local Zones, see Available Regions in the Amazon Elastic Compute Cloud User Guide.

To create a subnet in an Outpost, set this value to the Availability Zone for the Outpost and specify the Outpost ARN.

createSubnet_dryRun :: Lens' CreateSubnet (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

createSubnet_cidrBlock :: Lens' CreateSubnet Text

The IPv4 network range for the subnet, in CIDR notation. For example, 10.0.0.0/24. We modify the specified CIDR block to its canonical form; for example, if you specify 100.68.0.18/18, we modify it to 100.68.0.0/18.

createSubnet_vpcId :: Lens' CreateSubnet Text

The ID of the VPC.

Destructuring the Response

data CreateSubnetResponse

See: newCreateSubnetResponse smart constructor.

Constructors

CreateSubnetResponse'

Fields

subnet :: Maybe Subnet

Information about the subnet.

httpStatus :: Int

The response's http status code.

newCreateSubnetResponse

Arguments

:: Int
$sel:httpStatus:CreateSubnetResponse'

-> CreateSubnetResponse

Create a value of CreateSubnetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:subnet:CreateSubnetResponse', createSubnetResponse_subnet - Information about the subnet.

$sel:httpStatus:CreateSubnetResponse', createSubnetResponse_httpStatus - The response's http status code.

Response Lenses

createSubnetResponse_subnet :: Lens' CreateSubnetResponse (Maybe Subnet)

Information about the subnet.

createSubnetResponse_httpStatus :: Lens' CreateSubnetResponse Int

The response's http status code.