Network.AWS.EC2.CreateTrafficMirrorFilter

Description

Creates a Traffic Mirror filter.

A Traffic Mirror filter is a set of rules that defines the traffic to mirror.

By default, no traffic is mirrored. To mirror traffic, use CreateTrafficMirrorFilterRule to add Traffic Mirror rules to the filter. The rules you add define what traffic gets mirrored. You can also use ModifyTrafficMirrorFilterNetworkServices to mirror supported network services.

Creating a Request

data CreateTrafficMirrorFilter

See: newCreateTrafficMirrorFilter smart constructor.

Constructors

CreateTrafficMirrorFilter'

Fields

tagSpecifications :: Maybe [TagSpecification]

The tags to assign to a Traffic Mirror filter.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

description :: Maybe Text

The description of the Traffic Mirror filter.

clientToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to Ensure Idempotency.

newCreateTrafficMirrorFilter :: CreateTrafficMirrorFilter

Create a value of CreateTrafficMirrorFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tagSpecifications:CreateTrafficMirrorFilter', createTrafficMirrorFilter_tagSpecifications - The tags to assign to a Traffic Mirror filter.

$sel:dryRun:CreateTrafficMirrorFilter', createTrafficMirrorFilter_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:description:CreateTrafficMirrorFilter', createTrafficMirrorFilter_description - The description of the Traffic Mirror filter.

$sel:clientToken:CreateTrafficMirrorFilter', createTrafficMirrorFilter_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to Ensure Idempotency.

Request Lenses

createTrafficMirrorFilter_tagSpecifications :: Lens' CreateTrafficMirrorFilter (Maybe [TagSpecification])

The tags to assign to a Traffic Mirror filter.

createTrafficMirrorFilter_dryRun :: Lens' CreateTrafficMirrorFilter (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

createTrafficMirrorFilter_description :: Lens' CreateTrafficMirrorFilter (Maybe Text)

The description of the Traffic Mirror filter.

createTrafficMirrorFilter_clientToken :: Lens' CreateTrafficMirrorFilter (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to Ensure Idempotency.

Destructuring the Response

data CreateTrafficMirrorFilterResponse

See: newCreateTrafficMirrorFilterResponse smart constructor.

Constructors

CreateTrafficMirrorFilterResponse'

Fields

trafficMirrorFilter :: Maybe TrafficMirrorFilter

Information about the Traffic Mirror filter.

clientToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to Ensure Idempotency.

httpStatus :: Int

The response's http status code.

newCreateTrafficMirrorFilterResponse

Arguments

:: Int
$sel:httpStatus:CreateTrafficMirrorFilterResponse'

-> CreateTrafficMirrorFilterResponse

Create a value of CreateTrafficMirrorFilterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:trafficMirrorFilter:CreateTrafficMirrorFilterResponse', createTrafficMirrorFilterResponse_trafficMirrorFilter - Information about the Traffic Mirror filter.

$sel:clientToken:CreateTrafficMirrorFilter', createTrafficMirrorFilterResponse_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to Ensure Idempotency.

$sel:httpStatus:CreateTrafficMirrorFilterResponse', createTrafficMirrorFilterResponse_httpStatus - The response's http status code.

Response Lenses

createTrafficMirrorFilterResponse_trafficMirrorFilter :: Lens' CreateTrafficMirrorFilterResponse (Maybe TrafficMirrorFilter)

Information about the Traffic Mirror filter.

createTrafficMirrorFilterResponse_clientToken :: Lens' CreateTrafficMirrorFilterResponse (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to Ensure Idempotency.

createTrafficMirrorFilterResponse_httpStatus :: Lens' CreateTrafficMirrorFilterResponse Int

The response's http status code.