Network.AWS.EC2.CreateTrafficMirrorFilterRule

Description

Creates a Traffic Mirror filter rule.

A Traffic Mirror rule defines the Traffic Mirror source traffic to mirror.

You need the Traffic Mirror filter ID when you create the rule.

Creating a Request

data CreateTrafficMirrorFilterRule

See: newCreateTrafficMirrorFilterRule smart constructor.

Constructors

CreateTrafficMirrorFilterRule'

Fields

clientToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to ensure idempotency.

protocol :: Maybe Int

The protocol, for example UDP, to assign to the Traffic Mirror rule.

For information about the protocol value, see Protocol Numbers on the Internet Assigned Numbers Authority (IANA) website.

destinationPortRange :: Maybe TrafficMirrorPortRangeRequest

The destination port range.

sourcePortRange :: Maybe TrafficMirrorPortRangeRequest

The source port range.

description :: Maybe Text

The description of the Traffic Mirror rule.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

trafficMirrorFilterId :: Text

The ID of the filter that this rule is associated with.

trafficDirection :: TrafficDirection

The type of traffic.

ruleNumber :: Int

The number of the Traffic Mirror rule. This number must be unique for each Traffic Mirror rule in a given direction. The rules are processed in ascending order by rule number.

ruleAction :: TrafficMirrorRuleAction

The action to take on the filtered traffic.

destinationCidrBlock :: Text

The destination CIDR block to assign to the Traffic Mirror rule.

sourceCidrBlock :: Text

The source CIDR block to assign to the Traffic Mirror rule.

newCreateTrafficMirrorFilterRule

Arguments

:: Text
$sel:trafficMirrorFilterId:CreateTrafficMirrorFilterRule'

-> TrafficDirection
$sel:trafficDirection:CreateTrafficMirrorFilterRule'

-> Int
$sel:ruleNumber:CreateTrafficMirrorFilterRule'

-> TrafficMirrorRuleAction
$sel:ruleAction:CreateTrafficMirrorFilterRule'

-> Text
$sel:destinationCidrBlock:CreateTrafficMirrorFilterRule'

-> Text
$sel:sourceCidrBlock:CreateTrafficMirrorFilterRule'

-> CreateTrafficMirrorFilterRule

Create a value of CreateTrafficMirrorFilterRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateTrafficMirrorFilterRule', createTrafficMirrorFilterRule_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to ensure idempotency.

$sel:protocol:CreateTrafficMirrorFilterRule', createTrafficMirrorFilterRule_protocol - The protocol, for example UDP, to assign to the Traffic Mirror rule.

For information about the protocol value, see Protocol Numbers on the Internet Assigned Numbers Authority (IANA) website.

$sel:destinationPortRange:CreateTrafficMirrorFilterRule', createTrafficMirrorFilterRule_destinationPortRange - The destination port range.

$sel:sourcePortRange:CreateTrafficMirrorFilterRule', createTrafficMirrorFilterRule_sourcePortRange - The source port range.

$sel:description:CreateTrafficMirrorFilterRule', createTrafficMirrorFilterRule_description - The description of the Traffic Mirror rule.

$sel:dryRun:CreateTrafficMirrorFilterRule', createTrafficMirrorFilterRule_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:trafficMirrorFilterId:CreateTrafficMirrorFilterRule', createTrafficMirrorFilterRule_trafficMirrorFilterId - The ID of the filter that this rule is associated with.

$sel:trafficDirection:CreateTrafficMirrorFilterRule', createTrafficMirrorFilterRule_trafficDirection - The type of traffic.

$sel:ruleNumber:CreateTrafficMirrorFilterRule', createTrafficMirrorFilterRule_ruleNumber - The number of the Traffic Mirror rule. This number must be unique for each Traffic Mirror rule in a given direction. The rules are processed in ascending order by rule number.

$sel:ruleAction:CreateTrafficMirrorFilterRule', createTrafficMirrorFilterRule_ruleAction - The action to take on the filtered traffic.

$sel:destinationCidrBlock:CreateTrafficMirrorFilterRule', createTrafficMirrorFilterRule_destinationCidrBlock - The destination CIDR block to assign to the Traffic Mirror rule.

$sel:sourceCidrBlock:CreateTrafficMirrorFilterRule', createTrafficMirrorFilterRule_sourceCidrBlock - The source CIDR block to assign to the Traffic Mirror rule.

Request Lenses

createTrafficMirrorFilterRule_clientToken :: Lens' CreateTrafficMirrorFilterRule (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to ensure idempotency.

createTrafficMirrorFilterRule_protocol :: Lens' CreateTrafficMirrorFilterRule (Maybe Int)

The protocol, for example UDP, to assign to the Traffic Mirror rule.

For information about the protocol value, see Protocol Numbers on the Internet Assigned Numbers Authority (IANA) website.

createTrafficMirrorFilterRule_destinationPortRange :: Lens' CreateTrafficMirrorFilterRule (Maybe TrafficMirrorPortRangeRequest)

The destination port range.

createTrafficMirrorFilterRule_sourcePortRange :: Lens' CreateTrafficMirrorFilterRule (Maybe TrafficMirrorPortRangeRequest)

The source port range.

createTrafficMirrorFilterRule_description :: Lens' CreateTrafficMirrorFilterRule (Maybe Text)

The description of the Traffic Mirror rule.

createTrafficMirrorFilterRule_dryRun :: Lens' CreateTrafficMirrorFilterRule (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

createTrafficMirrorFilterRule_trafficMirrorFilterId :: Lens' CreateTrafficMirrorFilterRule Text

The ID of the filter that this rule is associated with.

createTrafficMirrorFilterRule_trafficDirection :: Lens' CreateTrafficMirrorFilterRule TrafficDirection

The type of traffic.

createTrafficMirrorFilterRule_ruleNumber :: Lens' CreateTrafficMirrorFilterRule Int

The number of the Traffic Mirror rule. This number must be unique for each Traffic Mirror rule in a given direction. The rules are processed in ascending order by rule number.

createTrafficMirrorFilterRule_ruleAction :: Lens' CreateTrafficMirrorFilterRule TrafficMirrorRuleAction

The action to take on the filtered traffic.

createTrafficMirrorFilterRule_destinationCidrBlock :: Lens' CreateTrafficMirrorFilterRule Text

The destination CIDR block to assign to the Traffic Mirror rule.

createTrafficMirrorFilterRule_sourceCidrBlock :: Lens' CreateTrafficMirrorFilterRule Text

The source CIDR block to assign to the Traffic Mirror rule.

Destructuring the Response

data CreateTrafficMirrorFilterRuleResponse

See: newCreateTrafficMirrorFilterRuleResponse smart constructor.

Constructors

CreateTrafficMirrorFilterRuleResponse'

Fields

trafficMirrorFilterRule :: Maybe TrafficMirrorFilterRule

The Traffic Mirror rule.

clientToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to ensure idempotency.

httpStatus :: Int

The response's http status code.

newCreateTrafficMirrorFilterRuleResponse

Arguments

:: Int
$sel:httpStatus:CreateTrafficMirrorFilterRuleResponse'

-> CreateTrafficMirrorFilterRuleResponse

Create a value of CreateTrafficMirrorFilterRuleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:trafficMirrorFilterRule:CreateTrafficMirrorFilterRuleResponse', createTrafficMirrorFilterRuleResponse_trafficMirrorFilterRule - The Traffic Mirror rule.

$sel:clientToken:CreateTrafficMirrorFilterRule', createTrafficMirrorFilterRuleResponse_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to ensure idempotency.

$sel:httpStatus:CreateTrafficMirrorFilterRuleResponse', createTrafficMirrorFilterRuleResponse_httpStatus - The response's http status code.

Response Lenses

createTrafficMirrorFilterRuleResponse_trafficMirrorFilterRule :: Lens' CreateTrafficMirrorFilterRuleResponse (Maybe TrafficMirrorFilterRule)

The Traffic Mirror rule.

createTrafficMirrorFilterRuleResponse_clientToken :: Lens' CreateTrafficMirrorFilterRuleResponse (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to ensure idempotency.

createTrafficMirrorFilterRuleResponse_httpStatus :: Lens' CreateTrafficMirrorFilterRuleResponse Int

The response's http status code.