Network.AWS.EC2.CreateTrafficMirrorSession

Description

Creates a Traffic Mirror session.

A Traffic Mirror session actively copies packets from a Traffic Mirror source to a Traffic Mirror target. Create a filter, and then assign it to the session to define a subset of the traffic to mirror, for example all TCP traffic.

The Traffic Mirror source and the Traffic Mirror target (monitoring appliances) can be in the same VPC, or in a different VPC connected via VPC peering or a transit gateway.

By default, no traffic is mirrored. Use CreateTrafficMirrorFilter to create filter rules that specify the traffic to mirror.

Creating a Request

data CreateTrafficMirrorSession

See: newCreateTrafficMirrorSession smart constructor.

Constructors

CreateTrafficMirrorSession'

Fields

clientToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to ensure idempotency.

packetLength :: Maybe Int

The number of bytes in each packet to mirror. These are bytes after the VXLAN header. Do not specify this parameter when you want to mirror the entire packet. To mirror a subset of the packet, set this to the length (in bytes) that you want to mirror. For example, if you set this value to 100, then the first 100 bytes that meet the filter criteria are copied to the target.

If you do not want to mirror the entire packet, use the PacketLength parameter to specify the number of bytes in each packet to mirror.

tagSpecifications :: Maybe [TagSpecification]

The tags to assign to a Traffic Mirror session.

virtualNetworkId :: Maybe Int

The VXLAN ID for the Traffic Mirror session. For more information about the VXLAN protocol, see RFC 7348. If you do not specify a VirtualNetworkId, an account-wide unique id is chosen at random.

description :: Maybe Text

The description of the Traffic Mirror session.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

networkInterfaceId :: Text

The ID of the source network interface.

trafficMirrorTargetId :: Text

The ID of the Traffic Mirror target.

trafficMirrorFilterId :: Text

The ID of the Traffic Mirror filter.

sessionNumber :: Int

The session number determines the order in which sessions are evaluated when an interface is used by multiple sessions. The first session with a matching filter is the one that mirrors the packets.

Valid values are 1-32766.

newCreateTrafficMirrorSession

Arguments

:: Text
$sel:networkInterfaceId:CreateTrafficMirrorSession'

-> Text
$sel:trafficMirrorTargetId:CreateTrafficMirrorSession'

-> Text
$sel:trafficMirrorFilterId:CreateTrafficMirrorSession'

-> Int
$sel:sessionNumber:CreateTrafficMirrorSession'

-> CreateTrafficMirrorSession

Create a value of CreateTrafficMirrorSession with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateTrafficMirrorSession', createTrafficMirrorSession_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to ensure idempotency.

$sel:packetLength:CreateTrafficMirrorSession', createTrafficMirrorSession_packetLength - The number of bytes in each packet to mirror. These are bytes after the VXLAN header. Do not specify this parameter when you want to mirror the entire packet. To mirror a subset of the packet, set this to the length (in bytes) that you want to mirror. For example, if you set this value to 100, then the first 100 bytes that meet the filter criteria are copied to the target.

If you do not want to mirror the entire packet, use the PacketLength parameter to specify the number of bytes in each packet to mirror.

$sel:tagSpecifications:CreateTrafficMirrorSession', createTrafficMirrorSession_tagSpecifications - The tags to assign to a Traffic Mirror session.

$sel:virtualNetworkId:CreateTrafficMirrorSession', createTrafficMirrorSession_virtualNetworkId - The VXLAN ID for the Traffic Mirror session. For more information about the VXLAN protocol, see RFC 7348. If you do not specify a VirtualNetworkId, an account-wide unique id is chosen at random.

$sel:description:CreateTrafficMirrorSession', createTrafficMirrorSession_description - The description of the Traffic Mirror session.

$sel:dryRun:CreateTrafficMirrorSession', createTrafficMirrorSession_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:networkInterfaceId:CreateTrafficMirrorSession', createTrafficMirrorSession_networkInterfaceId - The ID of the source network interface.

$sel:trafficMirrorTargetId:CreateTrafficMirrorSession', createTrafficMirrorSession_trafficMirrorTargetId - The ID of the Traffic Mirror target.

$sel:trafficMirrorFilterId:CreateTrafficMirrorSession', createTrafficMirrorSession_trafficMirrorFilterId - The ID of the Traffic Mirror filter.

$sel:sessionNumber:CreateTrafficMirrorSession', createTrafficMirrorSession_sessionNumber - The session number determines the order in which sessions are evaluated when an interface is used by multiple sessions. The first session with a matching filter is the one that mirrors the packets.

Valid values are 1-32766.

Request Lenses

createTrafficMirrorSession_clientToken :: Lens' CreateTrafficMirrorSession (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to ensure idempotency.

createTrafficMirrorSession_packetLength :: Lens' CreateTrafficMirrorSession (Maybe Int)

The number of bytes in each packet to mirror. These are bytes after the VXLAN header. Do not specify this parameter when you want to mirror the entire packet. To mirror a subset of the packet, set this to the length (in bytes) that you want to mirror. For example, if you set this value to 100, then the first 100 bytes that meet the filter criteria are copied to the target.

If you do not want to mirror the entire packet, use the PacketLength parameter to specify the number of bytes in each packet to mirror.

createTrafficMirrorSession_tagSpecifications :: Lens' CreateTrafficMirrorSession (Maybe [TagSpecification])

The tags to assign to a Traffic Mirror session.

createTrafficMirrorSession_virtualNetworkId :: Lens' CreateTrafficMirrorSession (Maybe Int)

The VXLAN ID for the Traffic Mirror session. For more information about the VXLAN protocol, see RFC 7348. If you do not specify a VirtualNetworkId, an account-wide unique id is chosen at random.

createTrafficMirrorSession_description :: Lens' CreateTrafficMirrorSession (Maybe Text)

The description of the Traffic Mirror session.

createTrafficMirrorSession_dryRun :: Lens' CreateTrafficMirrorSession (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

createTrafficMirrorSession_networkInterfaceId :: Lens' CreateTrafficMirrorSession Text

The ID of the source network interface.

createTrafficMirrorSession_trafficMirrorTargetId :: Lens' CreateTrafficMirrorSession Text

The ID of the Traffic Mirror target.

createTrafficMirrorSession_trafficMirrorFilterId :: Lens' CreateTrafficMirrorSession Text

The ID of the Traffic Mirror filter.

createTrafficMirrorSession_sessionNumber :: Lens' CreateTrafficMirrorSession Int

The session number determines the order in which sessions are evaluated when an interface is used by multiple sessions. The first session with a matching filter is the one that mirrors the packets.

Valid values are 1-32766.

Destructuring the Response

data CreateTrafficMirrorSessionResponse

See: newCreateTrafficMirrorSessionResponse smart constructor.

Constructors

CreateTrafficMirrorSessionResponse'

Fields

trafficMirrorSession :: Maybe TrafficMirrorSession

Information about the Traffic Mirror session.

clientToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to ensure idempotency.

httpStatus :: Int

The response's http status code.

newCreateTrafficMirrorSessionResponse

Arguments

:: Int
$sel:httpStatus:CreateTrafficMirrorSessionResponse'

-> CreateTrafficMirrorSessionResponse

Create a value of CreateTrafficMirrorSessionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:trafficMirrorSession:CreateTrafficMirrorSessionResponse', createTrafficMirrorSessionResponse_trafficMirrorSession - Information about the Traffic Mirror session.

$sel:clientToken:CreateTrafficMirrorSession', createTrafficMirrorSessionResponse_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to ensure idempotency.

$sel:httpStatus:CreateTrafficMirrorSessionResponse', createTrafficMirrorSessionResponse_httpStatus - The response's http status code.

Response Lenses

createTrafficMirrorSessionResponse_trafficMirrorSession :: Lens' CreateTrafficMirrorSessionResponse (Maybe TrafficMirrorSession)

Information about the Traffic Mirror session.

createTrafficMirrorSessionResponse_clientToken :: Lens' CreateTrafficMirrorSessionResponse (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to ensure idempotency.

createTrafficMirrorSessionResponse_httpStatus :: Lens' CreateTrafficMirrorSessionResponse Int

The response's http status code.