Network.AWS.EC2.CreateTrafficMirrorTarget

Description

Creates a target for your Traffic Mirror session.

A Traffic Mirror target is the destination for mirrored traffic. The Traffic Mirror source and the Traffic Mirror target (monitoring appliances) can be in the same VPC, or in different VPCs connected via VPC peering or a transit gateway.

A Traffic Mirror target can be a network interface, or a Network Load Balancer.

To use the target in a Traffic Mirror session, use CreateTrafficMirrorSession.

Creating a Request

data CreateTrafficMirrorTarget

See: newCreateTrafficMirrorTarget smart constructor.

Constructors

CreateTrafficMirrorTarget'

Fields

tagSpecifications :: Maybe [TagSpecification]

The tags to assign to the Traffic Mirror target.

networkLoadBalancerArn :: Maybe Text

The Amazon Resource Name (ARN) of the Network Load Balancer that is associated with the target.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

description :: Maybe Text

The description of the Traffic Mirror target.

networkInterfaceId :: Maybe Text

The network interface ID that is associated with the target.

clientToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to Ensure Idempotency.

newCreateTrafficMirrorTarget :: CreateTrafficMirrorTarget

Create a value of CreateTrafficMirrorTarget with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tagSpecifications:CreateTrafficMirrorTarget', createTrafficMirrorTarget_tagSpecifications - The tags to assign to the Traffic Mirror target.

$sel:networkLoadBalancerArn:CreateTrafficMirrorTarget', createTrafficMirrorTarget_networkLoadBalancerArn - The Amazon Resource Name (ARN) of the Network Load Balancer that is associated with the target.

$sel:dryRun:CreateTrafficMirrorTarget', createTrafficMirrorTarget_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:description:CreateTrafficMirrorTarget', createTrafficMirrorTarget_description - The description of the Traffic Mirror target.

$sel:networkInterfaceId:CreateTrafficMirrorTarget', createTrafficMirrorTarget_networkInterfaceId - The network interface ID that is associated with the target.

$sel:clientToken:CreateTrafficMirrorTarget', createTrafficMirrorTarget_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to Ensure Idempotency.

Request Lenses

createTrafficMirrorTarget_tagSpecifications :: Lens' CreateTrafficMirrorTarget (Maybe [TagSpecification])

The tags to assign to the Traffic Mirror target.

createTrafficMirrorTarget_networkLoadBalancerArn :: Lens' CreateTrafficMirrorTarget (Maybe Text)

The Amazon Resource Name (ARN) of the Network Load Balancer that is associated with the target.

createTrafficMirrorTarget_dryRun :: Lens' CreateTrafficMirrorTarget (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

createTrafficMirrorTarget_description :: Lens' CreateTrafficMirrorTarget (Maybe Text)

The description of the Traffic Mirror target.

createTrafficMirrorTarget_networkInterfaceId :: Lens' CreateTrafficMirrorTarget (Maybe Text)

The network interface ID that is associated with the target.

createTrafficMirrorTarget_clientToken :: Lens' CreateTrafficMirrorTarget (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to Ensure Idempotency.

Destructuring the Response

data CreateTrafficMirrorTargetResponse

See: newCreateTrafficMirrorTargetResponse smart constructor.

Constructors

CreateTrafficMirrorTargetResponse'

Fields

trafficMirrorTarget :: Maybe TrafficMirrorTarget

Information about the Traffic Mirror target.

clientToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to Ensure Idempotency.

httpStatus :: Int

The response's http status code.

newCreateTrafficMirrorTargetResponse

Arguments

:: Int
$sel:httpStatus:CreateTrafficMirrorTargetResponse'

-> CreateTrafficMirrorTargetResponse

Create a value of CreateTrafficMirrorTargetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:trafficMirrorTarget:CreateTrafficMirrorTargetResponse', createTrafficMirrorTargetResponse_trafficMirrorTarget - Information about the Traffic Mirror target.

$sel:clientToken:CreateTrafficMirrorTarget', createTrafficMirrorTargetResponse_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to Ensure Idempotency.

$sel:httpStatus:CreateTrafficMirrorTargetResponse', createTrafficMirrorTargetResponse_httpStatus - The response's http status code.

Response Lenses

createTrafficMirrorTargetResponse_trafficMirrorTarget :: Lens' CreateTrafficMirrorTargetResponse (Maybe TrafficMirrorTarget)

Information about the Traffic Mirror target.

createTrafficMirrorTargetResponse_clientToken :: Lens' CreateTrafficMirrorTargetResponse (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to Ensure Idempotency.

createTrafficMirrorTargetResponse_httpStatus :: Lens' CreateTrafficMirrorTargetResponse Int

The response's http status code.