Network.AWS.EC2.CreateTransitGatewayMulticastDomain

Description

Creates a multicast domain using the specified transit gateway.

The transit gateway must be in the available state before you create a domain. Use DescribeTransitGateways to see the state of transit gateway.

Creating a Request

data CreateTransitGatewayMulticastDomain

See: newCreateTransitGatewayMulticastDomain smart constructor.

Constructors

CreateTransitGatewayMulticastDomain'

Fields

tagSpecifications :: Maybe [TagSpecification]

The tags for the transit gateway multicast domain.

options :: Maybe CreateTransitGatewayMulticastDomainRequestOptions

The options for the transit gateway multicast domain.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

transitGatewayId :: Text

The ID of the transit gateway.

newCreateTransitGatewayMulticastDomain

Arguments

:: Text
$sel:transitGatewayId:CreateTransitGatewayMulticastDomain'

-> CreateTransitGatewayMulticastDomain

Create a value of CreateTransitGatewayMulticastDomain with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tagSpecifications:CreateTransitGatewayMulticastDomain', createTransitGatewayMulticastDomain_tagSpecifications - The tags for the transit gateway multicast domain.

$sel:options:CreateTransitGatewayMulticastDomain', createTransitGatewayMulticastDomain_options - The options for the transit gateway multicast domain.

$sel:dryRun:CreateTransitGatewayMulticastDomain', createTransitGatewayMulticastDomain_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:transitGatewayId:CreateTransitGatewayMulticastDomain', createTransitGatewayMulticastDomain_transitGatewayId - The ID of the transit gateway.

Request Lenses

createTransitGatewayMulticastDomain_tagSpecifications :: Lens' CreateTransitGatewayMulticastDomain (Maybe [TagSpecification])

The tags for the transit gateway multicast domain.

createTransitGatewayMulticastDomain_options :: Lens' CreateTransitGatewayMulticastDomain (Maybe CreateTransitGatewayMulticastDomainRequestOptions)

The options for the transit gateway multicast domain.

createTransitGatewayMulticastDomain_dryRun :: Lens' CreateTransitGatewayMulticastDomain (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

createTransitGatewayMulticastDomain_transitGatewayId :: Lens' CreateTransitGatewayMulticastDomain Text

The ID of the transit gateway.

Destructuring the Response

data CreateTransitGatewayMulticastDomainResponse

See: newCreateTransitGatewayMulticastDomainResponse smart constructor.

Constructors

CreateTransitGatewayMulticastDomainResponse'

Fields

transitGatewayMulticastDomain :: Maybe TransitGatewayMulticastDomain

Information about the transit gateway multicast domain.

httpStatus :: Int

The response's http status code.

newCreateTransitGatewayMulticastDomainResponse

Arguments

:: Int
$sel:httpStatus:CreateTransitGatewayMulticastDomainResponse'

-> CreateTransitGatewayMulticastDomainResponse

Create a value of CreateTransitGatewayMulticastDomainResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:transitGatewayMulticastDomain:CreateTransitGatewayMulticastDomainResponse', createTransitGatewayMulticastDomainResponse_transitGatewayMulticastDomain - Information about the transit gateway multicast domain.

$sel:httpStatus:CreateTransitGatewayMulticastDomainResponse', createTransitGatewayMulticastDomainResponse_httpStatus - The response's http status code.

Response Lenses

createTransitGatewayMulticastDomainResponse_transitGatewayMulticastDomain :: Lens' CreateTransitGatewayMulticastDomainResponse (Maybe TransitGatewayMulticastDomain)

Information about the transit gateway multicast domain.

createTransitGatewayMulticastDomainResponse_httpStatus :: Lens' CreateTransitGatewayMulticastDomainResponse Int

The response's http status code.