Network.AWS.EC2.CreateTransitGatewayPeeringAttachment

Description

Requests a transit gateway peering attachment between the specified transit gateway (requester) and a peer transit gateway (accepter). The transit gateways must be in different Regions. The peer transit gateway can be in your account or a different Amazon Web Services account.

After you create the peering attachment, the owner of the accepter transit gateway must accept the attachment request.

Creating a Request

data CreateTransitGatewayPeeringAttachment

See: newCreateTransitGatewayPeeringAttachment smart constructor.

Constructors

CreateTransitGatewayPeeringAttachment'

Fields

tagSpecifications :: Maybe [TagSpecification]

The tags to apply to the transit gateway peering attachment.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

transitGatewayId :: Text

The ID of the transit gateway.

peerTransitGatewayId :: Text

The ID of the peer transit gateway with which to create the peering attachment.

peerAccountId :: Text

The ID of the Amazon Web Services account that owns the peer transit gateway.

peerRegion :: Text

The Region where the peer transit gateway is located.

newCreateTransitGatewayPeeringAttachment

Arguments

:: Text
$sel:transitGatewayId:CreateTransitGatewayPeeringAttachment'

-> Text
$sel:peerTransitGatewayId:CreateTransitGatewayPeeringAttachment'

-> Text
$sel:peerAccountId:CreateTransitGatewayPeeringAttachment'

-> Text
$sel:peerRegion:CreateTransitGatewayPeeringAttachment'

-> CreateTransitGatewayPeeringAttachment

Create a value of CreateTransitGatewayPeeringAttachment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tagSpecifications:CreateTransitGatewayPeeringAttachment', createTransitGatewayPeeringAttachment_tagSpecifications - The tags to apply to the transit gateway peering attachment.

$sel:dryRun:CreateTransitGatewayPeeringAttachment', createTransitGatewayPeeringAttachment_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:transitGatewayId:CreateTransitGatewayPeeringAttachment', createTransitGatewayPeeringAttachment_transitGatewayId - The ID of the transit gateway.

$sel:peerTransitGatewayId:CreateTransitGatewayPeeringAttachment', createTransitGatewayPeeringAttachment_peerTransitGatewayId - The ID of the peer transit gateway with which to create the peering attachment.

$sel:peerAccountId:CreateTransitGatewayPeeringAttachment', createTransitGatewayPeeringAttachment_peerAccountId - The ID of the Amazon Web Services account that owns the peer transit gateway.

$sel:peerRegion:CreateTransitGatewayPeeringAttachment', createTransitGatewayPeeringAttachment_peerRegion - The Region where the peer transit gateway is located.

Request Lenses

createTransitGatewayPeeringAttachment_tagSpecifications :: Lens' CreateTransitGatewayPeeringAttachment (Maybe [TagSpecification])

The tags to apply to the transit gateway peering attachment.

createTransitGatewayPeeringAttachment_dryRun :: Lens' CreateTransitGatewayPeeringAttachment (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

createTransitGatewayPeeringAttachment_transitGatewayId :: Lens' CreateTransitGatewayPeeringAttachment Text

The ID of the transit gateway.

createTransitGatewayPeeringAttachment_peerTransitGatewayId :: Lens' CreateTransitGatewayPeeringAttachment Text

The ID of the peer transit gateway with which to create the peering attachment.

createTransitGatewayPeeringAttachment_peerAccountId :: Lens' CreateTransitGatewayPeeringAttachment Text

The ID of the Amazon Web Services account that owns the peer transit gateway.

createTransitGatewayPeeringAttachment_peerRegion :: Lens' CreateTransitGatewayPeeringAttachment Text

The Region where the peer transit gateway is located.

Destructuring the Response

data CreateTransitGatewayPeeringAttachmentResponse

See: newCreateTransitGatewayPeeringAttachmentResponse smart constructor.

Constructors

CreateTransitGatewayPeeringAttachmentResponse'

Fields

transitGatewayPeeringAttachment :: Maybe TransitGatewayPeeringAttachment

The transit gateway peering attachment.

httpStatus :: Int

The response's http status code.

newCreateTransitGatewayPeeringAttachmentResponse

Arguments

:: Int
$sel:httpStatus:CreateTransitGatewayPeeringAttachmentResponse'

-> CreateTransitGatewayPeeringAttachmentResponse

Create a value of CreateTransitGatewayPeeringAttachmentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:transitGatewayPeeringAttachment:CreateTransitGatewayPeeringAttachmentResponse', createTransitGatewayPeeringAttachmentResponse_transitGatewayPeeringAttachment - The transit gateway peering attachment.

$sel:httpStatus:CreateTransitGatewayPeeringAttachmentResponse', createTransitGatewayPeeringAttachmentResponse_httpStatus - The response's http status code.

Response Lenses

createTransitGatewayPeeringAttachmentResponse_transitGatewayPeeringAttachment :: Lens' CreateTransitGatewayPeeringAttachmentResponse (Maybe TransitGatewayPeeringAttachment)

The transit gateway peering attachment.

createTransitGatewayPeeringAttachmentResponse_httpStatus :: Lens' CreateTransitGatewayPeeringAttachmentResponse Int

The response's http status code.