Network.AWS.EC2.CreateTransitGatewayPrefixListReference

Description

Creates a reference (route) to a prefix list in a specified transit gateway route table.

Creating a Request

data CreateTransitGatewayPrefixListReference

See: newCreateTransitGatewayPrefixListReference smart constructor.

Constructors

CreateTransitGatewayPrefixListReference'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

blackhole :: Maybe Bool

Indicates whether to drop traffic that matches this route.

transitGatewayAttachmentId :: Maybe Text

The ID of the attachment to which traffic is routed.

transitGatewayRouteTableId :: Text

The ID of the transit gateway route table.

prefixListId :: Text

The ID of the prefix list that is used for destination matches.

newCreateTransitGatewayPrefixListReference

Arguments

:: Text
$sel:transitGatewayRouteTableId:CreateTransitGatewayPrefixListReference'

-> Text
$sel:prefixListId:CreateTransitGatewayPrefixListReference'

-> CreateTransitGatewayPrefixListReference

Create a value of CreateTransitGatewayPrefixListReference with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:CreateTransitGatewayPrefixListReference', createTransitGatewayPrefixListReference_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:blackhole:CreateTransitGatewayPrefixListReference', createTransitGatewayPrefixListReference_blackhole - Indicates whether to drop traffic that matches this route.

$sel:transitGatewayAttachmentId:CreateTransitGatewayPrefixListReference', createTransitGatewayPrefixListReference_transitGatewayAttachmentId - The ID of the attachment to which traffic is routed.

$sel:transitGatewayRouteTableId:CreateTransitGatewayPrefixListReference', createTransitGatewayPrefixListReference_transitGatewayRouteTableId - The ID of the transit gateway route table.

$sel:prefixListId:CreateTransitGatewayPrefixListReference', createTransitGatewayPrefixListReference_prefixListId - The ID of the prefix list that is used for destination matches.

Request Lenses

createTransitGatewayPrefixListReference_dryRun :: Lens' CreateTransitGatewayPrefixListReference (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

createTransitGatewayPrefixListReference_blackhole :: Lens' CreateTransitGatewayPrefixListReference (Maybe Bool)

Indicates whether to drop traffic that matches this route.

createTransitGatewayPrefixListReference_transitGatewayAttachmentId :: Lens' CreateTransitGatewayPrefixListReference (Maybe Text)

The ID of the attachment to which traffic is routed.

createTransitGatewayPrefixListReference_transitGatewayRouteTableId :: Lens' CreateTransitGatewayPrefixListReference Text

The ID of the transit gateway route table.

createTransitGatewayPrefixListReference_prefixListId :: Lens' CreateTransitGatewayPrefixListReference Text

The ID of the prefix list that is used for destination matches.

Destructuring the Response

data CreateTransitGatewayPrefixListReferenceResponse

See: newCreateTransitGatewayPrefixListReferenceResponse smart constructor.

Constructors

CreateTransitGatewayPrefixListReferenceResponse'

Fields

transitGatewayPrefixListReference :: Maybe TransitGatewayPrefixListReference

Information about the prefix list reference.

httpStatus :: Int

The response's http status code.

newCreateTransitGatewayPrefixListReferenceResponse

Arguments

:: Int
$sel:httpStatus:CreateTransitGatewayPrefixListReferenceResponse'

-> CreateTransitGatewayPrefixListReferenceResponse

Create a value of CreateTransitGatewayPrefixListReferenceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:transitGatewayPrefixListReference:CreateTransitGatewayPrefixListReferenceResponse', createTransitGatewayPrefixListReferenceResponse_transitGatewayPrefixListReference - Information about the prefix list reference.

$sel:httpStatus:CreateTransitGatewayPrefixListReferenceResponse', createTransitGatewayPrefixListReferenceResponse_httpStatus - The response's http status code.

Response Lenses

createTransitGatewayPrefixListReferenceResponse_transitGatewayPrefixListReference :: Lens' CreateTransitGatewayPrefixListReferenceResponse (Maybe TransitGatewayPrefixListReference)

Information about the prefix list reference.

createTransitGatewayPrefixListReferenceResponse_httpStatus :: Lens' CreateTransitGatewayPrefixListReferenceResponse Int

The response's http status code.