Network.AWS.EC2.CreateTransitGatewayVpcAttachment

Description

Attaches the specified VPC to the specified transit gateway.

If you attach a VPC with a CIDR range that overlaps the CIDR range of a VPC that is already attached, the new VPC CIDR range is not propagated to the default propagation route table.

To send VPC traffic to an attached transit gateway, add a route to the VPC route table using CreateRoute.

Creating a Request

data CreateTransitGatewayVpcAttachment

See: newCreateTransitGatewayVpcAttachment smart constructor.

Constructors

CreateTransitGatewayVpcAttachment'

Fields

tagSpecifications :: Maybe [TagSpecification]

The tags to apply to the VPC attachment.

options :: Maybe CreateTransitGatewayVpcAttachmentRequestOptions

The VPC attachment options.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

transitGatewayId :: Text

The ID of the transit gateway.

vpcId :: Text

The ID of the VPC.

subnetIds :: [Text]

The IDs of one or more subnets. You can specify only one subnet per Availability Zone. You must specify at least one subnet, but we recommend that you specify two subnets for better availability. The transit gateway uses one IP address from each specified subnet.

newCreateTransitGatewayVpcAttachment

Arguments

:: Text
$sel:transitGatewayId:CreateTransitGatewayVpcAttachment'

-> Text
$sel:vpcId:CreateTransitGatewayVpcAttachment'

-> CreateTransitGatewayVpcAttachment

Create a value of CreateTransitGatewayVpcAttachment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tagSpecifications:CreateTransitGatewayVpcAttachment', createTransitGatewayVpcAttachment_tagSpecifications - The tags to apply to the VPC attachment.

$sel:options:CreateTransitGatewayVpcAttachment', createTransitGatewayVpcAttachment_options - The VPC attachment options.

$sel:dryRun:CreateTransitGatewayVpcAttachment', createTransitGatewayVpcAttachment_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:transitGatewayId:CreateTransitGatewayVpcAttachment', createTransitGatewayVpcAttachment_transitGatewayId - The ID of the transit gateway.

$sel:vpcId:CreateTransitGatewayVpcAttachment', createTransitGatewayVpcAttachment_vpcId - The ID of the VPC.

$sel:subnetIds:CreateTransitGatewayVpcAttachment', createTransitGatewayVpcAttachment_subnetIds - The IDs of one or more subnets. You can specify only one subnet per Availability Zone. You must specify at least one subnet, but we recommend that you specify two subnets for better availability. The transit gateway uses one IP address from each specified subnet.

Request Lenses

createTransitGatewayVpcAttachment_tagSpecifications :: Lens' CreateTransitGatewayVpcAttachment (Maybe [TagSpecification])

The tags to apply to the VPC attachment.

createTransitGatewayVpcAttachment_options :: Lens' CreateTransitGatewayVpcAttachment (Maybe CreateTransitGatewayVpcAttachmentRequestOptions)

The VPC attachment options.

createTransitGatewayVpcAttachment_dryRun :: Lens' CreateTransitGatewayVpcAttachment (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

createTransitGatewayVpcAttachment_transitGatewayId :: Lens' CreateTransitGatewayVpcAttachment Text

The ID of the transit gateway.

createTransitGatewayVpcAttachment_vpcId :: Lens' CreateTransitGatewayVpcAttachment Text

The ID of the VPC.

createTransitGatewayVpcAttachment_subnetIds :: Lens' CreateTransitGatewayVpcAttachment [Text]

The IDs of one or more subnets. You can specify only one subnet per Availability Zone. You must specify at least one subnet, but we recommend that you specify two subnets for better availability. The transit gateway uses one IP address from each specified subnet.

Destructuring the Response

data CreateTransitGatewayVpcAttachmentResponse

See: newCreateTransitGatewayVpcAttachmentResponse smart constructor.

Constructors

CreateTransitGatewayVpcAttachmentResponse'

Fields

transitGatewayVpcAttachment :: Maybe TransitGatewayVpcAttachment

Information about the VPC attachment.

httpStatus :: Int

The response's http status code.

newCreateTransitGatewayVpcAttachmentResponse

Arguments

:: Int
$sel:httpStatus:CreateTransitGatewayVpcAttachmentResponse'

-> CreateTransitGatewayVpcAttachmentResponse

Create a value of CreateTransitGatewayVpcAttachmentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:transitGatewayVpcAttachment:CreateTransitGatewayVpcAttachmentResponse', createTransitGatewayVpcAttachmentResponse_transitGatewayVpcAttachment - Information about the VPC attachment.

$sel:httpStatus:CreateTransitGatewayVpcAttachmentResponse', createTransitGatewayVpcAttachmentResponse_httpStatus - The response's http status code.

Response Lenses

createTransitGatewayVpcAttachmentResponse_transitGatewayVpcAttachment :: Lens' CreateTransitGatewayVpcAttachmentResponse (Maybe TransitGatewayVpcAttachment)

Information about the VPC attachment.

createTransitGatewayVpcAttachmentResponse_httpStatus :: Lens' CreateTransitGatewayVpcAttachmentResponse Int

The response's http status code.