Network.AWS.EC2.CreateVpc

Description

Creates a VPC with the specified IPv4 CIDR block. The smallest VPC you can create uses a /28 netmask (16 IPv4 addresses), and the largest uses a /16 netmask (65,536 IPv4 addresses). For more information about how large to make your VPC, see Your VPC and subnets in the Amazon Virtual Private Cloud User Guide.

You can optionally request an IPv6 CIDR block for the VPC. You can request an Amazon-provided IPv6 CIDR block from Amazon's pool of IPv6 addresses, or an IPv6 CIDR block from an IPv6 address pool that you provisioned through bring your own IP addresses (BYOIP).

By default, each instance you launch in the VPC has the default DHCP options, which include only a default DNS server that we provide (AmazonProvidedDNS). For more information, see DHCP options sets in the Amazon Virtual Private Cloud User Guide.

You can specify the instance tenancy value for the VPC when you create it. You can't change this value for the VPC after you create it. For more information, see Dedicated Instances in the Amazon Elastic Compute Cloud User Guide.

Creating a Request

data CreateVpc

See: newCreateVpc smart constructor.

Constructors

CreateVpc'

Fields

tagSpecifications :: Maybe [TagSpecification]

The tags to assign to the VPC.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

instanceTenancy :: Maybe Tenancy

The tenancy options for instances launched into the VPC. For default, instances are launched with shared tenancy by default. You can launch instances with any tenancy into a shared tenancy VPC. For dedicated, instances are launched as dedicated tenancy instances by default. You can only launch instances with a tenancy of dedicated or host into a dedicated tenancy VPC.

Important: The host value cannot be used with this parameter. Use the default or dedicated values only.

Default: default

ipv6Pool :: Maybe Text

The ID of an IPv6 address pool from which to allocate the IPv6 CIDR block.

ipv6CidrBlock :: Maybe Text

The IPv6 CIDR block from the IPv6 address pool. You must also specify Ipv6Pool in the request.

To let Amazon choose the IPv6 CIDR block for you, omit this parameter.

ipv6CidrBlockNetworkBorderGroup :: Maybe Text

The name of the location from which we advertise the IPV6 CIDR block. Use this parameter to limit the address to this location.

You must set AmazonProvidedIpv6CidrBlock to true to use this parameter.

amazonProvidedIpv6CidrBlock :: Maybe Bool

Requests an Amazon-provided IPv6 CIDR block with a /56 prefix length for the VPC. You cannot specify the range of IP addresses, or the size of the CIDR block.

cidrBlock :: Text

The IPv4 network range for the VPC, in CIDR notation. For example, 10.0.0.0/16. We modify the specified CIDR block to its canonical form; for example, if you specify 100.68.0.18/18, we modify it to 100.68.0.0/18.

newCreateVpc

Arguments

:: Text
$sel:cidrBlock:CreateVpc'

-> CreateVpc

Create a value of CreateVpc with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tagSpecifications:CreateVpc', createVpc_tagSpecifications - The tags to assign to the VPC.

$sel:dryRun:CreateVpc', createVpc_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:instanceTenancy:CreateVpc', createVpc_instanceTenancy - The tenancy options for instances launched into the VPC. For default, instances are launched with shared tenancy by default. You can launch instances with any tenancy into a shared tenancy VPC. For dedicated, instances are launched as dedicated tenancy instances by default. You can only launch instances with a tenancy of dedicated or host into a dedicated tenancy VPC.

Important: The host value cannot be used with this parameter. Use the default or dedicated values only.

Default: default

$sel:ipv6Pool:CreateVpc', createVpc_ipv6Pool - The ID of an IPv6 address pool from which to allocate the IPv6 CIDR block.

$sel:ipv6CidrBlock:CreateVpc', createVpc_ipv6CidrBlock - The IPv6 CIDR block from the IPv6 address pool. You must also specify Ipv6Pool in the request.

To let Amazon choose the IPv6 CIDR block for you, omit this parameter.

$sel:ipv6CidrBlockNetworkBorderGroup:CreateVpc', createVpc_ipv6CidrBlockNetworkBorderGroup - The name of the location from which we advertise the IPV6 CIDR block. Use this parameter to limit the address to this location.

You must set AmazonProvidedIpv6CidrBlock to true to use this parameter.

$sel:amazonProvidedIpv6CidrBlock:CreateVpc', createVpc_amazonProvidedIpv6CidrBlock - Requests an Amazon-provided IPv6 CIDR block with a /56 prefix length for the VPC. You cannot specify the range of IP addresses, or the size of the CIDR block.

$sel:cidrBlock:CreateVpc', createVpc_cidrBlock - The IPv4 network range for the VPC, in CIDR notation. For example, 10.0.0.0/16. We modify the specified CIDR block to its canonical form; for example, if you specify 100.68.0.18/18, we modify it to 100.68.0.0/18.

Request Lenses

createVpc_tagSpecifications :: Lens' CreateVpc (Maybe [TagSpecification])

The tags to assign to the VPC.

createVpc_dryRun :: Lens' CreateVpc (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

createVpc_instanceTenancy :: Lens' CreateVpc (Maybe Tenancy)

The tenancy options for instances launched into the VPC. For default, instances are launched with shared tenancy by default. You can launch instances with any tenancy into a shared tenancy VPC. For dedicated, instances are launched as dedicated tenancy instances by default. You can only launch instances with a tenancy of dedicated or host into a dedicated tenancy VPC.

Important: The host value cannot be used with this parameter. Use the default or dedicated values only.

Default: default

createVpc_ipv6Pool :: Lens' CreateVpc (Maybe Text)

The ID of an IPv6 address pool from which to allocate the IPv6 CIDR block.

createVpc_ipv6CidrBlock :: Lens' CreateVpc (Maybe Text)

The IPv6 CIDR block from the IPv6 address pool. You must also specify Ipv6Pool in the request.

To let Amazon choose the IPv6 CIDR block for you, omit this parameter.

createVpc_ipv6CidrBlockNetworkBorderGroup :: Lens' CreateVpc (Maybe Text)

The name of the location from which we advertise the IPV6 CIDR block. Use this parameter to limit the address to this location.

You must set AmazonProvidedIpv6CidrBlock to true to use this parameter.

createVpc_amazonProvidedIpv6CidrBlock :: Lens' CreateVpc (Maybe Bool)

Requests an Amazon-provided IPv6 CIDR block with a /56 prefix length for the VPC. You cannot specify the range of IP addresses, or the size of the CIDR block.

createVpc_cidrBlock :: Lens' CreateVpc Text

The IPv4 network range for the VPC, in CIDR notation. For example, 10.0.0.0/16. We modify the specified CIDR block to its canonical form; for example, if you specify 100.68.0.18/18, we modify it to 100.68.0.0/18.

Destructuring the Response

data CreateVpcResponse

See: newCreateVpcResponse smart constructor.

Constructors

CreateVpcResponse'

Fields

vpc :: Maybe Vpc

Information about the VPC.

httpStatus :: Int

The response's http status code.

newCreateVpcResponse

Arguments

:: Int
$sel:httpStatus:CreateVpcResponse'

-> CreateVpcResponse

Create a value of CreateVpcResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpc:CreateVpcResponse', createVpcResponse_vpc - Information about the VPC.

$sel:httpStatus:CreateVpcResponse', createVpcResponse_httpStatus - The response's http status code.

Response Lenses

createVpcResponse_vpc :: Lens' CreateVpcResponse (Maybe Vpc)

Information about the VPC.

createVpcResponse_httpStatus :: Lens' CreateVpcResponse Int

The response's http status code.