Network.AWS.EC2.CreateVpcEndpoint

Description

Creates a VPC endpoint for a specified service. An endpoint enables you to create a private connection between your VPC and the service. The service may be provided by AWS, an AWS Marketplace Partner, or another AWS account. For more information, see VPC Endpoints in the Amazon Virtual Private Cloud User Guide.

A gateway endpoint serves as a target for a route in your route table for traffic destined for the AWS service. You can specify an endpoint policy to attach to the endpoint, which will control access to the service from your VPC. You can also specify the VPC route tables that use the endpoint.

An interface endpoint is a network interface in your subnet that serves as an endpoint for communicating with the specified service. You can specify the subnets in which to create an endpoint, and the security groups to associate with the endpoint network interface.

A GatewayLoadBalancer endpoint is a network interface in your subnet that serves an endpoint for communicating with a Gateway Load Balancer that you've configured as a VPC endpoint service.

Use DescribeVpcEndpointServices to get a list of supported services.

Creating a Request

data CreateVpcEndpoint

Contains the parameters for CreateVpcEndpoint.

See: newCreateVpcEndpoint smart constructor.

Constructors

CreateVpcEndpoint'

Fields

policyDocument :: Maybe Text

(Interface and gateway endpoints) A policy to attach to the endpoint that controls access to the service. The policy must be in valid JSON format. If this parameter is not specified, we attach a default policy that allows full access to the service.

securityGroupIds :: Maybe [Text]

(Interface endpoint) The ID of one or more security groups to associate with the endpoint network interface.

tagSpecifications :: Maybe [TagSpecification]

The tags to associate with the endpoint.

routeTableIds :: Maybe [Text]

(Gateway endpoint) One or more route table IDs.

vpcEndpointType :: Maybe VpcEndpointType

The type of endpoint.

Default: Gateway

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

subnetIds :: Maybe [Text]

(Interface and Gateway Load Balancer endpoints) The ID of one or more subnets in which to create an endpoint network interface. For a Gateway Load Balancer endpoint, you can specify one subnet only.

privateDnsEnabled :: Maybe Bool

(Interface endpoint) Indicates whether to associate a private hosted zone with the specified VPC. The private hosted zone contains a record set for the default public DNS name for the service for the Region (for example, kinesis.us-east-1.amazonaws.com), which resolves to the private IP addresses of the endpoint network interfaces in the VPC. This enables you to make requests to the default public DNS name for the service instead of the public DNS names that are automatically generated by the VPC endpoint service.

To use a private hosted zone, you must set the following VPC attributes to true: enableDnsHostnames and enableDnsSupport. Use ModifyVpcAttribute to set the VPC attributes.

Default: true

clientToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to Ensure Idempotency.

vpcId :: Text

The ID of the VPC in which the endpoint will be used.

serviceName :: Text

The service name. To get a list of available services, use the DescribeVpcEndpointServices request, or get the name from the service provider.

newCreateVpcEndpoint

Arguments

:: Text
$sel:vpcId:CreateVpcEndpoint'

-> Text
$sel:serviceName:CreateVpcEndpoint'

-> CreateVpcEndpoint

Create a value of CreateVpcEndpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policyDocument:CreateVpcEndpoint', createVpcEndpoint_policyDocument - (Interface and gateway endpoints) A policy to attach to the endpoint that controls access to the service. The policy must be in valid JSON format. If this parameter is not specified, we attach a default policy that allows full access to the service.

$sel:securityGroupIds:CreateVpcEndpoint', createVpcEndpoint_securityGroupIds - (Interface endpoint) The ID of one or more security groups to associate with the endpoint network interface.

$sel:tagSpecifications:CreateVpcEndpoint', createVpcEndpoint_tagSpecifications - The tags to associate with the endpoint.

$sel:routeTableIds:CreateVpcEndpoint', createVpcEndpoint_routeTableIds - (Gateway endpoint) One or more route table IDs.

$sel:vpcEndpointType:CreateVpcEndpoint', createVpcEndpoint_vpcEndpointType - The type of endpoint.

Default: Gateway

$sel:dryRun:CreateVpcEndpoint', createVpcEndpoint_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:subnetIds:CreateVpcEndpoint', createVpcEndpoint_subnetIds - (Interface and Gateway Load Balancer endpoints) The ID of one or more subnets in which to create an endpoint network interface. For a Gateway Load Balancer endpoint, you can specify one subnet only.

$sel:privateDnsEnabled:CreateVpcEndpoint', createVpcEndpoint_privateDnsEnabled - (Interface endpoint) Indicates whether to associate a private hosted zone with the specified VPC. The private hosted zone contains a record set for the default public DNS name for the service for the Region (for example, kinesis.us-east-1.amazonaws.com), which resolves to the private IP addresses of the endpoint network interfaces in the VPC. This enables you to make requests to the default public DNS name for the service instead of the public DNS names that are automatically generated by the VPC endpoint service.

To use a private hosted zone, you must set the following VPC attributes to true: enableDnsHostnames and enableDnsSupport. Use ModifyVpcAttribute to set the VPC attributes.

Default: true

$sel:clientToken:CreateVpcEndpoint', createVpcEndpoint_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to Ensure Idempotency.

$sel:vpcId:CreateVpcEndpoint', createVpcEndpoint_vpcId - The ID of the VPC in which the endpoint will be used.

$sel:serviceName:CreateVpcEndpoint', createVpcEndpoint_serviceName - The service name. To get a list of available services, use the DescribeVpcEndpointServices request, or get the name from the service provider.

Request Lenses

createVpcEndpoint_policyDocument :: Lens' CreateVpcEndpoint (Maybe Text)

(Interface and gateway endpoints) A policy to attach to the endpoint that controls access to the service. The policy must be in valid JSON format. If this parameter is not specified, we attach a default policy that allows full access to the service.

createVpcEndpoint_securityGroupIds :: Lens' CreateVpcEndpoint (Maybe [Text])

(Interface endpoint) The ID of one or more security groups to associate with the endpoint network interface.

createVpcEndpoint_tagSpecifications :: Lens' CreateVpcEndpoint (Maybe [TagSpecification])

The tags to associate with the endpoint.

createVpcEndpoint_routeTableIds :: Lens' CreateVpcEndpoint (Maybe [Text])

(Gateway endpoint) One or more route table IDs.

createVpcEndpoint_vpcEndpointType :: Lens' CreateVpcEndpoint (Maybe VpcEndpointType)

The type of endpoint.

Default: Gateway

createVpcEndpoint_dryRun :: Lens' CreateVpcEndpoint (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

createVpcEndpoint_subnetIds :: Lens' CreateVpcEndpoint (Maybe [Text])

(Interface and Gateway Load Balancer endpoints) The ID of one or more subnets in which to create an endpoint network interface. For a Gateway Load Balancer endpoint, you can specify one subnet only.

createVpcEndpoint_privateDnsEnabled :: Lens' CreateVpcEndpoint (Maybe Bool)

(Interface endpoint) Indicates whether to associate a private hosted zone with the specified VPC. The private hosted zone contains a record set for the default public DNS name for the service for the Region (for example, kinesis.us-east-1.amazonaws.com), which resolves to the private IP addresses of the endpoint network interfaces in the VPC. This enables you to make requests to the default public DNS name for the service instead of the public DNS names that are automatically generated by the VPC endpoint service.

To use a private hosted zone, you must set the following VPC attributes to true: enableDnsHostnames and enableDnsSupport. Use ModifyVpcAttribute to set the VPC attributes.

Default: true

createVpcEndpoint_clientToken :: Lens' CreateVpcEndpoint (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to Ensure Idempotency.

createVpcEndpoint_vpcId :: Lens' CreateVpcEndpoint Text

The ID of the VPC in which the endpoint will be used.

createVpcEndpoint_serviceName :: Lens' CreateVpcEndpoint Text

The service name. To get a list of available services, use the DescribeVpcEndpointServices request, or get the name from the service provider.

Destructuring the Response

data CreateVpcEndpointResponse

Contains the output of CreateVpcEndpoint.

See: newCreateVpcEndpointResponse smart constructor.

Constructors

CreateVpcEndpointResponse'

Fields

vpcEndpoint :: Maybe VpcEndpoint

Information about the endpoint.

clientToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

httpStatus :: Int

The response's http status code.

newCreateVpcEndpointResponse

Arguments

:: Int
$sel:httpStatus:CreateVpcEndpointResponse'

-> CreateVpcEndpointResponse

Create a value of CreateVpcEndpointResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpcEndpoint:CreateVpcEndpointResponse', createVpcEndpointResponse_vpcEndpoint - Information about the endpoint.

$sel:clientToken:CreateVpcEndpoint', createVpcEndpointResponse_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

$sel:httpStatus:CreateVpcEndpointResponse', createVpcEndpointResponse_httpStatus - The response's http status code.

Response Lenses

createVpcEndpointResponse_vpcEndpoint :: Lens' CreateVpcEndpointResponse (Maybe VpcEndpoint)

Information about the endpoint.

createVpcEndpointResponse_clientToken :: Lens' CreateVpcEndpointResponse (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

createVpcEndpointResponse_httpStatus :: Lens' CreateVpcEndpointResponse Int

The response's http status code.