Network.AWS.EC2.CreateVpcEndpointConnectionNotification

Description

Creates a connection notification for a specified VPC endpoint or VPC endpoint service. A connection notification notifies you of specific endpoint events. You must create an SNS topic to receive notifications. For more information, see Create a Topic in the Amazon Simple Notification Service Developer Guide.

You can create a connection notification for interface endpoints only.

Creating a Request

data CreateVpcEndpointConnectionNotification

See: newCreateVpcEndpointConnectionNotification smart constructor.

Constructors

CreateVpcEndpointConnectionNotification'

Fields

clientToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to ensure idempotency.

serviceId :: Maybe Text

The ID of the endpoint service.

vpcEndpointId :: Maybe Text

The ID of the endpoint.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

connectionNotificationArn :: Text

The ARN of the SNS topic for the notifications.

connectionEvents :: [Text]

One or more endpoint events for which to receive notifications. Valid values are Accept, Connect, Delete, and Reject.

newCreateVpcEndpointConnectionNotification

Arguments

:: Text
$sel:connectionNotificationArn:CreateVpcEndpointConnectionNotification'

-> CreateVpcEndpointConnectionNotification

Create a value of CreateVpcEndpointConnectionNotification with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateVpcEndpointConnectionNotification', createVpcEndpointConnectionNotification_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to ensure idempotency.

$sel:serviceId:CreateVpcEndpointConnectionNotification', createVpcEndpointConnectionNotification_serviceId - The ID of the endpoint service.

$sel:vpcEndpointId:CreateVpcEndpointConnectionNotification', createVpcEndpointConnectionNotification_vpcEndpointId - The ID of the endpoint.

$sel:dryRun:CreateVpcEndpointConnectionNotification', createVpcEndpointConnectionNotification_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:connectionNotificationArn:CreateVpcEndpointConnectionNotification', createVpcEndpointConnectionNotification_connectionNotificationArn - The ARN of the SNS topic for the notifications.

$sel:connectionEvents:CreateVpcEndpointConnectionNotification', createVpcEndpointConnectionNotification_connectionEvents - One or more endpoint events for which to receive notifications. Valid values are Accept, Connect, Delete, and Reject.

Request Lenses

createVpcEndpointConnectionNotification_clientToken :: Lens' CreateVpcEndpointConnectionNotification (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to ensure idempotency.

createVpcEndpointConnectionNotification_serviceId :: Lens' CreateVpcEndpointConnectionNotification (Maybe Text)

The ID of the endpoint service.

createVpcEndpointConnectionNotification_vpcEndpointId :: Lens' CreateVpcEndpointConnectionNotification (Maybe Text)

The ID of the endpoint.

createVpcEndpointConnectionNotification_dryRun :: Lens' CreateVpcEndpointConnectionNotification (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

createVpcEndpointConnectionNotification_connectionNotificationArn :: Lens' CreateVpcEndpointConnectionNotification Text

The ARN of the SNS topic for the notifications.

createVpcEndpointConnectionNotification_connectionEvents :: Lens' CreateVpcEndpointConnectionNotification [Text]

One or more endpoint events for which to receive notifications. Valid values are Accept, Connect, Delete, and Reject.

Destructuring the Response

data CreateVpcEndpointConnectionNotificationResponse

See: newCreateVpcEndpointConnectionNotificationResponse smart constructor.

Constructors

CreateVpcEndpointConnectionNotificationResponse'

Fields

clientToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

connectionNotification :: Maybe ConnectionNotification

Information about the notification.

httpStatus :: Int

The response's http status code.

newCreateVpcEndpointConnectionNotificationResponse

Arguments

:: Int
$sel:httpStatus:CreateVpcEndpointConnectionNotificationResponse'

-> CreateVpcEndpointConnectionNotificationResponse

Create a value of CreateVpcEndpointConnectionNotificationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateVpcEndpointConnectionNotification', createVpcEndpointConnectionNotificationResponse_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

$sel:connectionNotification:CreateVpcEndpointConnectionNotificationResponse', createVpcEndpointConnectionNotificationResponse_connectionNotification - Information about the notification.

$sel:httpStatus:CreateVpcEndpointConnectionNotificationResponse', createVpcEndpointConnectionNotificationResponse_httpStatus - The response's http status code.

Response Lenses

createVpcEndpointConnectionNotificationResponse_clientToken :: Lens' CreateVpcEndpointConnectionNotificationResponse (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

createVpcEndpointConnectionNotificationResponse_connectionNotification :: Lens' CreateVpcEndpointConnectionNotificationResponse (Maybe ConnectionNotification)

Information about the notification.

createVpcEndpointConnectionNotificationResponse_httpStatus :: Lens' CreateVpcEndpointConnectionNotificationResponse Int

The response's http status code.