Network.AWS.EC2.CreateVpnConnection

Description

Creates a VPN connection between an existing virtual private gateway or transit gateway and a customer gateway. The supported connection type is ipsec.1.

The response includes information that you need to give to your network administrator to configure your customer gateway.

We strongly recommend that you use HTTPS when calling this operation because the response contains sensitive cryptographic information for configuring your customer gateway device.

If you decide to shut down your VPN connection for any reason and later create a new VPN connection, you must reconfigure your customer gateway with the new information returned from this call.

This is an idempotent operation. If you perform the operation more than once, Amazon EC2 doesn't return an error.

For more information, see Amazon Web Services Site-to-Site VPN in the Amazon Web Services Site-to-Site VPN User Guide.

Creating a Request

data CreateVpnConnection

Contains the parameters for CreateVpnConnection.

See: newCreateVpnConnection smart constructor.

Constructors

CreateVpnConnection'

Fields

tagSpecifications :: Maybe [TagSpecification]

The tags to apply to the VPN connection.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

options :: Maybe VpnConnectionOptionsSpecification

The options for the VPN connection.

vpnGatewayId :: Maybe Text

The ID of the virtual private gateway. If you specify a virtual private gateway, you cannot specify a transit gateway.

transitGatewayId :: Maybe Text

The ID of the transit gateway. If you specify a transit gateway, you cannot specify a virtual private gateway.

customerGatewayId :: Text

The ID of the customer gateway.

type' :: Text

The type of VPN connection (ipsec.1).

newCreateVpnConnection

Arguments

:: Text
$sel:customerGatewayId:CreateVpnConnection'

-> Text
$sel:type':CreateVpnConnection'

-> CreateVpnConnection

Create a value of CreateVpnConnection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tagSpecifications:CreateVpnConnection', createVpnConnection_tagSpecifications - The tags to apply to the VPN connection.

$sel:dryRun:CreateVpnConnection', createVpnConnection_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:options:CreateVpnConnection', createVpnConnection_options - The options for the VPN connection.

$sel:vpnGatewayId:CreateVpnConnection', createVpnConnection_vpnGatewayId - The ID of the virtual private gateway. If you specify a virtual private gateway, you cannot specify a transit gateway.

$sel:transitGatewayId:CreateVpnConnection', createVpnConnection_transitGatewayId - The ID of the transit gateway. If you specify a transit gateway, you cannot specify a virtual private gateway.

$sel:customerGatewayId:CreateVpnConnection', createVpnConnection_customerGatewayId - The ID of the customer gateway.

$sel:type':CreateVpnConnection', createVpnConnection_type - The type of VPN connection (ipsec.1).

Request Lenses

createVpnConnection_tagSpecifications :: Lens' CreateVpnConnection (Maybe [TagSpecification])

The tags to apply to the VPN connection.

createVpnConnection_dryRun :: Lens' CreateVpnConnection (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

createVpnConnection_options :: Lens' CreateVpnConnection (Maybe VpnConnectionOptionsSpecification)

The options for the VPN connection.

createVpnConnection_vpnGatewayId :: Lens' CreateVpnConnection (Maybe Text)

The ID of the virtual private gateway. If you specify a virtual private gateway, you cannot specify a transit gateway.

createVpnConnection_transitGatewayId :: Lens' CreateVpnConnection (Maybe Text)

The ID of the transit gateway. If you specify a transit gateway, you cannot specify a virtual private gateway.

createVpnConnection_customerGatewayId :: Lens' CreateVpnConnection Text

The ID of the customer gateway.

createVpnConnection_type :: Lens' CreateVpnConnection Text

The type of VPN connection (ipsec.1).

Destructuring the Response

data CreateVpnConnectionResponse

Contains the output of CreateVpnConnection.

See: newCreateVpnConnectionResponse smart constructor.

Constructors

CreateVpnConnectionResponse'

Fields

vpnConnection :: Maybe VpnConnection

Information about the VPN connection.

httpStatus :: Int

The response's http status code.

newCreateVpnConnectionResponse

Arguments

:: Int
$sel:httpStatus:CreateVpnConnectionResponse'

-> CreateVpnConnectionResponse

Create a value of CreateVpnConnectionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpnConnection:CreateVpnConnectionResponse', createVpnConnectionResponse_vpnConnection - Information about the VPN connection.

$sel:httpStatus:CreateVpnConnectionResponse', createVpnConnectionResponse_httpStatus - The response's http status code.

Response Lenses

createVpnConnectionResponse_vpnConnection :: Lens' CreateVpnConnectionResponse (Maybe VpnConnection)

Information about the VPN connection.

createVpnConnectionResponse_httpStatus :: Lens' CreateVpnConnectionResponse Int

The response's http status code.