Network.AWS.EC2.DeleteInternetGateway

Description

Deletes the specified internet gateway. You must detach the internet gateway from the VPC before you can delete it.

Creating a Request

data DeleteInternetGateway

See: newDeleteInternetGateway smart constructor.

Constructors

DeleteInternetGateway'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

internetGatewayId :: Text

The ID of the internet gateway.

newDeleteInternetGateway

Arguments

:: Text
$sel:internetGatewayId:DeleteInternetGateway'

-> DeleteInternetGateway

Create a value of DeleteInternetGateway with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:DeleteInternetGateway', deleteInternetGateway_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:internetGatewayId:DeleteInternetGateway', deleteInternetGateway_internetGatewayId - The ID of the internet gateway.

Request Lenses

deleteInternetGateway_dryRun :: Lens' DeleteInternetGateway (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

deleteInternetGateway_internetGatewayId :: Lens' DeleteInternetGateway Text

The ID of the internet gateway.

Destructuring the Response

data DeleteInternetGatewayResponse

See: newDeleteInternetGatewayResponse smart constructor.

Constructors

DeleteInternetGatewayResponse'

newDeleteInternetGatewayResponse :: DeleteInternetGatewayResponse

Create a value of DeleteInternetGatewayResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.