Network.AWS.EC2.DeleteNetworkInsightsPath

Description

Deletes the specified path.

Creating a Request

data DeleteNetworkInsightsPath

See: newDeleteNetworkInsightsPath smart constructor.

Constructors

DeleteNetworkInsightsPath'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

networkInsightsPathId :: Text

The ID of the path.

newDeleteNetworkInsightsPath

Arguments

:: Text
$sel:networkInsightsPathId:DeleteNetworkInsightsPath'

-> DeleteNetworkInsightsPath

Create a value of DeleteNetworkInsightsPath with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:DeleteNetworkInsightsPath', deleteNetworkInsightsPath_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:networkInsightsPathId:DeleteNetworkInsightsPath', deleteNetworkInsightsPath_networkInsightsPathId - The ID of the path.

Request Lenses

deleteNetworkInsightsPath_dryRun :: Lens' DeleteNetworkInsightsPath (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

deleteNetworkInsightsPath_networkInsightsPathId :: Lens' DeleteNetworkInsightsPath Text

The ID of the path.

Destructuring the Response

data DeleteNetworkInsightsPathResponse

See: newDeleteNetworkInsightsPathResponse smart constructor.

Constructors

DeleteNetworkInsightsPathResponse'

Fields

networkInsightsPathId :: Maybe Text

The ID of the path.

httpStatus :: Int

The response's http status code.

newDeleteNetworkInsightsPathResponse

Arguments

:: Int
$sel:httpStatus:DeleteNetworkInsightsPathResponse'

-> DeleteNetworkInsightsPathResponse

Create a value of DeleteNetworkInsightsPathResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:networkInsightsPathId:DeleteNetworkInsightsPath', deleteNetworkInsightsPathResponse_networkInsightsPathId - The ID of the path.

$sel:httpStatus:DeleteNetworkInsightsPathResponse', deleteNetworkInsightsPathResponse_httpStatus - The response's http status code.

Response Lenses

deleteNetworkInsightsPathResponse_networkInsightsPathId :: Lens' DeleteNetworkInsightsPathResponse (Maybe Text)

The ID of the path.

deleteNetworkInsightsPathResponse_httpStatus :: Lens' DeleteNetworkInsightsPathResponse Int

The response's http status code.