Network.AWS.EC2.DeleteQueuedReservedInstances

Description

Deletes the queued purchases for the specified Reserved Instances.

Creating a Request

data DeleteQueuedReservedInstances

See: newDeleteQueuedReservedInstances smart constructor.

Constructors

DeleteQueuedReservedInstances'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

reservedInstancesIds :: NonEmpty Text

The IDs of the Reserved Instances.

newDeleteQueuedReservedInstances

Arguments

:: NonEmpty Text
$sel:reservedInstancesIds:DeleteQueuedReservedInstances'

-> DeleteQueuedReservedInstances

Create a value of DeleteQueuedReservedInstances with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:DeleteQueuedReservedInstances', deleteQueuedReservedInstances_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:reservedInstancesIds:DeleteQueuedReservedInstances', deleteQueuedReservedInstances_reservedInstancesIds - The IDs of the Reserved Instances.

Request Lenses

deleteQueuedReservedInstances_dryRun :: Lens' DeleteQueuedReservedInstances (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

deleteQueuedReservedInstances_reservedInstancesIds :: Lens' DeleteQueuedReservedInstances (NonEmpty Text)

The IDs of the Reserved Instances.

Destructuring the Response

data DeleteQueuedReservedInstancesResponse

See: newDeleteQueuedReservedInstancesResponse smart constructor.

Constructors

DeleteQueuedReservedInstancesResponse'

Fields

successfulQueuedPurchaseDeletions :: Maybe [SuccessfulQueuedPurchaseDeletion]

Information about the queued purchases that were successfully deleted.

failedQueuedPurchaseDeletions :: Maybe [FailedQueuedPurchaseDeletion]

Information about the queued purchases that could not be deleted.

httpStatus :: Int

The response's http status code.

newDeleteQueuedReservedInstancesResponse

Arguments

:: Int
$sel:httpStatus:DeleteQueuedReservedInstancesResponse'

-> DeleteQueuedReservedInstancesResponse

Create a value of DeleteQueuedReservedInstancesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:successfulQueuedPurchaseDeletions:DeleteQueuedReservedInstancesResponse', deleteQueuedReservedInstancesResponse_successfulQueuedPurchaseDeletions - Information about the queued purchases that were successfully deleted.

$sel:failedQueuedPurchaseDeletions:DeleteQueuedReservedInstancesResponse', deleteQueuedReservedInstancesResponse_failedQueuedPurchaseDeletions - Information about the queued purchases that could not be deleted.

$sel:httpStatus:DeleteQueuedReservedInstancesResponse', deleteQueuedReservedInstancesResponse_httpStatus - The response's http status code.

Response Lenses

deleteQueuedReservedInstancesResponse_successfulQueuedPurchaseDeletions :: Lens' DeleteQueuedReservedInstancesResponse (Maybe [SuccessfulQueuedPurchaseDeletion])

Information about the queued purchases that were successfully deleted.

deleteQueuedReservedInstancesResponse_failedQueuedPurchaseDeletions :: Lens' DeleteQueuedReservedInstancesResponse (Maybe [FailedQueuedPurchaseDeletion])

Information about the queued purchases that could not be deleted.

deleteQueuedReservedInstancesResponse_httpStatus :: Lens' DeleteQueuedReservedInstancesResponse Int

The response's http status code.