Network.AWS.EC2.DeleteRoute

Description

Deletes the specified route from the specified route table.

Creating a Request

data DeleteRoute

See: newDeleteRoute smart constructor.

Constructors

DeleteRoute'

Fields

destinationIpv6CidrBlock :: Maybe Text

The IPv6 CIDR range for the route. The value you specify must match the CIDR for the route exactly.

destinationPrefixListId :: Maybe Text

The ID of the prefix list for the route.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

destinationCidrBlock :: Maybe Text

The IPv4 CIDR range for the route. The value you specify must match the CIDR for the route exactly.

routeTableId :: Text

The ID of the route table.

newDeleteRoute

Arguments

:: Text
$sel:routeTableId:DeleteRoute'

-> DeleteRoute

Create a value of DeleteRoute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:destinationIpv6CidrBlock:DeleteRoute', deleteRoute_destinationIpv6CidrBlock - The IPv6 CIDR range for the route. The value you specify must match the CIDR for the route exactly.

$sel:destinationPrefixListId:DeleteRoute', deleteRoute_destinationPrefixListId - The ID of the prefix list for the route.

$sel:dryRun:DeleteRoute', deleteRoute_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:destinationCidrBlock:DeleteRoute', deleteRoute_destinationCidrBlock - The IPv4 CIDR range for the route. The value you specify must match the CIDR for the route exactly.

$sel:routeTableId:DeleteRoute', deleteRoute_routeTableId - The ID of the route table.

Request Lenses

deleteRoute_destinationIpv6CidrBlock :: Lens' DeleteRoute (Maybe Text)

The IPv6 CIDR range for the route. The value you specify must match the CIDR for the route exactly.

deleteRoute_destinationPrefixListId :: Lens' DeleteRoute (Maybe Text)

The ID of the prefix list for the route.

deleteRoute_dryRun :: Lens' DeleteRoute (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

deleteRoute_destinationCidrBlock :: Lens' DeleteRoute (Maybe Text)

The IPv4 CIDR range for the route. The value you specify must match the CIDR for the route exactly.

deleteRoute_routeTableId :: Lens' DeleteRoute Text

The ID of the route table.

Destructuring the Response

data DeleteRouteResponse

See: newDeleteRouteResponse smart constructor.

Constructors

DeleteRouteResponse'

newDeleteRouteResponse :: DeleteRouteResponse

Create a value of DeleteRouteResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.