Network.AWS.EC2.DeleteSecurityGroup

Description

Deletes a security group.

If you attempt to delete a security group that is associated with an instance, or is referenced by another security group, the operation fails with InvalidGroup.InUse in EC2-Classic or DependencyViolation in EC2-VPC.

Creating a Request

data DeleteSecurityGroup

See: newDeleteSecurityGroup smart constructor.

Constructors

DeleteSecurityGroup'

Fields

groupId :: Maybe Text

The ID of the security group. Required for a nondefault VPC.

groupName :: Maybe Text

EC2-Classic, default VPC

The name of the security group. You can specify either the security group name or the security group ID.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

newDeleteSecurityGroup :: DeleteSecurityGroup

Create a value of DeleteSecurityGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groupId:DeleteSecurityGroup', deleteSecurityGroup_groupId - The ID of the security group. Required for a nondefault VPC.

$sel:groupName:DeleteSecurityGroup', deleteSecurityGroup_groupName - [EC2-Classic, default VPC] The name of the security group. You can specify either the security group name or the security group ID.

$sel:dryRun:DeleteSecurityGroup', deleteSecurityGroup_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

Request Lenses

deleteSecurityGroup_groupId :: Lens' DeleteSecurityGroup (Maybe Text)

The ID of the security group. Required for a nondefault VPC.

deleteSecurityGroup_groupName :: Lens' DeleteSecurityGroup (Maybe Text)

EC2-Classic, default VPC

The name of the security group. You can specify either the security group name or the security group ID.

deleteSecurityGroup_dryRun :: Lens' DeleteSecurityGroup (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

Destructuring the Response

data DeleteSecurityGroupResponse

See: newDeleteSecurityGroupResponse smart constructor.

Constructors

DeleteSecurityGroupResponse'

newDeleteSecurityGroupResponse :: DeleteSecurityGroupResponse

Create a value of DeleteSecurityGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.