Network.AWS.EC2.DeleteTransitGatewayVpcAttachment

Description

Deletes the specified VPC attachment.

Creating a Request

data DeleteTransitGatewayVpcAttachment

See: newDeleteTransitGatewayVpcAttachment smart constructor.

Constructors

DeleteTransitGatewayVpcAttachment'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

transitGatewayAttachmentId :: Text

The ID of the attachment.

newDeleteTransitGatewayVpcAttachment

Arguments

:: Text
$sel:transitGatewayAttachmentId:DeleteTransitGatewayVpcAttachment'

-> DeleteTransitGatewayVpcAttachment

Create a value of DeleteTransitGatewayVpcAttachment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:DeleteTransitGatewayVpcAttachment', deleteTransitGatewayVpcAttachment_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:transitGatewayAttachmentId:DeleteTransitGatewayVpcAttachment', deleteTransitGatewayVpcAttachment_transitGatewayAttachmentId - The ID of the attachment.

Request Lenses

deleteTransitGatewayVpcAttachment_dryRun :: Lens' DeleteTransitGatewayVpcAttachment (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

deleteTransitGatewayVpcAttachment_transitGatewayAttachmentId :: Lens' DeleteTransitGatewayVpcAttachment Text

The ID of the attachment.

Destructuring the Response

data DeleteTransitGatewayVpcAttachmentResponse

See: newDeleteTransitGatewayVpcAttachmentResponse smart constructor.

Constructors

DeleteTransitGatewayVpcAttachmentResponse'

Fields

transitGatewayVpcAttachment :: Maybe TransitGatewayVpcAttachment

Information about the deleted VPC attachment.

httpStatus :: Int

The response's http status code.

newDeleteTransitGatewayVpcAttachmentResponse

Arguments

:: Int
$sel:httpStatus:DeleteTransitGatewayVpcAttachmentResponse'

-> DeleteTransitGatewayVpcAttachmentResponse

Create a value of DeleteTransitGatewayVpcAttachmentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:transitGatewayVpcAttachment:DeleteTransitGatewayVpcAttachmentResponse', deleteTransitGatewayVpcAttachmentResponse_transitGatewayVpcAttachment - Information about the deleted VPC attachment.

$sel:httpStatus:DeleteTransitGatewayVpcAttachmentResponse', deleteTransitGatewayVpcAttachmentResponse_httpStatus - The response's http status code.

Response Lenses

deleteTransitGatewayVpcAttachmentResponse_transitGatewayVpcAttachment :: Lens' DeleteTransitGatewayVpcAttachmentResponse (Maybe TransitGatewayVpcAttachment)

Information about the deleted VPC attachment.

deleteTransitGatewayVpcAttachmentResponse_httpStatus :: Lens' DeleteTransitGatewayVpcAttachmentResponse Int

The response's http status code.