Network.AWS.EC2.DeleteVpnGateway

Description

Deletes the specified virtual private gateway. You must first detach the virtual private gateway from the VPC. Note that you don't need to delete the virtual private gateway if you plan to delete and recreate the VPN connection between your VPC and your network.

Creating a Request

data DeleteVpnGateway

Contains the parameters for DeleteVpnGateway.

See: newDeleteVpnGateway smart constructor.

Constructors

DeleteVpnGateway'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

vpnGatewayId :: Text

The ID of the virtual private gateway.

newDeleteVpnGateway

Arguments

:: Text
$sel:vpnGatewayId:DeleteVpnGateway'

-> DeleteVpnGateway

Create a value of DeleteVpnGateway with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:DeleteVpnGateway', deleteVpnGateway_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:vpnGatewayId:DeleteVpnGateway', deleteVpnGateway_vpnGatewayId - The ID of the virtual private gateway.

Request Lenses

deleteVpnGateway_dryRun :: Lens' DeleteVpnGateway (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

deleteVpnGateway_vpnGatewayId :: Lens' DeleteVpnGateway Text

The ID of the virtual private gateway.

Destructuring the Response

data DeleteVpnGatewayResponse

See: newDeleteVpnGatewayResponse smart constructor.

Constructors

DeleteVpnGatewayResponse'

newDeleteVpnGatewayResponse :: DeleteVpnGatewayResponse

Create a value of DeleteVpnGatewayResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.