Network.AWS.EC2.DeregisterInstanceEventNotificationAttributes

Description

c

Deregisters tag keys to prevent tags that have the specified tag keys from being included in scheduled event notifications for resources in the Region.

Creating a Request

data DeregisterInstanceEventNotificationAttributes

See: newDeregisterInstanceEventNotificationAttributes smart constructor.

Constructors

DeregisterInstanceEventNotificationAttributes'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

instanceTagAttribute :: Maybe DeregisterInstanceTagAttributeRequest

Information about the tag keys to deregister.

newDeregisterInstanceEventNotificationAttributes :: DeregisterInstanceEventNotificationAttributes

Create a value of DeregisterInstanceEventNotificationAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:DeregisterInstanceEventNotificationAttributes', deregisterInstanceEventNotificationAttributes_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:instanceTagAttribute:DeregisterInstanceEventNotificationAttributes', deregisterInstanceEventNotificationAttributes_instanceTagAttribute - Information about the tag keys to deregister.

Request Lenses

deregisterInstanceEventNotificationAttributes_dryRun :: Lens' DeregisterInstanceEventNotificationAttributes (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

deregisterInstanceEventNotificationAttributes_instanceTagAttribute :: Lens' DeregisterInstanceEventNotificationAttributes (Maybe DeregisterInstanceTagAttributeRequest)

Information about the tag keys to deregister.

Destructuring the Response

data DeregisterInstanceEventNotificationAttributesResponse

See: newDeregisterInstanceEventNotificationAttributesResponse smart constructor.

Constructors

DeregisterInstanceEventNotificationAttributesResponse'

Fields

instanceTagAttribute :: Maybe InstanceTagNotificationAttribute

The resulting set of tag keys.

httpStatus :: Int

The response's http status code.

newDeregisterInstanceEventNotificationAttributesResponse

Arguments

:: Int
$sel:httpStatus:DeregisterInstanceEventNotificationAttributesResponse'

-> DeregisterInstanceEventNotificationAttributesResponse

Create a value of DeregisterInstanceEventNotificationAttributesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceTagAttribute:DeregisterInstanceEventNotificationAttributes', deregisterInstanceEventNotificationAttributesResponse_instanceTagAttribute - The resulting set of tag keys.

$sel:httpStatus:DeregisterInstanceEventNotificationAttributesResponse', deregisterInstanceEventNotificationAttributesResponse_httpStatus - The response's http status code.

Response Lenses

deregisterInstanceEventNotificationAttributesResponse_instanceTagAttribute :: Lens' DeregisterInstanceEventNotificationAttributesResponse (Maybe InstanceTagNotificationAttribute)

The resulting set of tag keys.

deregisterInstanceEventNotificationAttributesResponse_httpStatus :: Lens' DeregisterInstanceEventNotificationAttributesResponse Int

The response's http status code.