Network.AWS.EC2.DeregisterTransitGatewayMulticastGroupMembers

Description

Deregisters the specified members (network interfaces) from the transit gateway multicast group.

Creating a Request

data DeregisterTransitGatewayMulticastGroupMembers

See: newDeregisterTransitGatewayMulticastGroupMembers smart constructor.

Constructors

DeregisterTransitGatewayMulticastGroupMembers'

Fields

transitGatewayMulticastDomainId :: Maybe Text

The ID of the transit gateway multicast domain.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

networkInterfaceIds :: Maybe [Text]

The IDs of the group members' network interfaces.

groupIpAddress :: Maybe Text

The IP address assigned to the transit gateway multicast group.

newDeregisterTransitGatewayMulticastGroupMembers :: DeregisterTransitGatewayMulticastGroupMembers

Create a value of DeregisterTransitGatewayMulticastGroupMembers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:transitGatewayMulticastDomainId:DeregisterTransitGatewayMulticastGroupMembers', deregisterTransitGatewayMulticastGroupMembers_transitGatewayMulticastDomainId - The ID of the transit gateway multicast domain.

$sel:dryRun:DeregisterTransitGatewayMulticastGroupMembers', deregisterTransitGatewayMulticastGroupMembers_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:networkInterfaceIds:DeregisterTransitGatewayMulticastGroupMembers', deregisterTransitGatewayMulticastGroupMembers_networkInterfaceIds - The IDs of the group members' network interfaces.

$sel:groupIpAddress:DeregisterTransitGatewayMulticastGroupMembers', deregisterTransitGatewayMulticastGroupMembers_groupIpAddress - The IP address assigned to the transit gateway multicast group.

Request Lenses

deregisterTransitGatewayMulticastGroupMembers_transitGatewayMulticastDomainId :: Lens' DeregisterTransitGatewayMulticastGroupMembers (Maybe Text)

The ID of the transit gateway multicast domain.

deregisterTransitGatewayMulticastGroupMembers_dryRun :: Lens' DeregisterTransitGatewayMulticastGroupMembers (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

deregisterTransitGatewayMulticastGroupMembers_networkInterfaceIds :: Lens' DeregisterTransitGatewayMulticastGroupMembers (Maybe [Text])

The IDs of the group members' network interfaces.

deregisterTransitGatewayMulticastGroupMembers_groupIpAddress :: Lens' DeregisterTransitGatewayMulticastGroupMembers (Maybe Text)

The IP address assigned to the transit gateway multicast group.

Destructuring the Response

data DeregisterTransitGatewayMulticastGroupMembersResponse

See: newDeregisterTransitGatewayMulticastGroupMembersResponse smart constructor.

Constructors

DeregisterTransitGatewayMulticastGroupMembersResponse'

Fields

deregisteredMulticastGroupMembers :: Maybe TransitGatewayMulticastDeregisteredGroupMembers

Information about the deregistered members.

httpStatus :: Int

The response's http status code.

newDeregisterTransitGatewayMulticastGroupMembersResponse

Arguments

:: Int
$sel:httpStatus:DeregisterTransitGatewayMulticastGroupMembersResponse'

-> DeregisterTransitGatewayMulticastGroupMembersResponse

Create a value of DeregisterTransitGatewayMulticastGroupMembersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deregisteredMulticastGroupMembers:DeregisterTransitGatewayMulticastGroupMembersResponse', deregisterTransitGatewayMulticastGroupMembersResponse_deregisteredMulticastGroupMembers - Information about the deregistered members.

$sel:httpStatus:DeregisterTransitGatewayMulticastGroupMembersResponse', deregisterTransitGatewayMulticastGroupMembersResponse_httpStatus - The response's http status code.

Response Lenses

deregisterTransitGatewayMulticastGroupMembersResponse_deregisteredMulticastGroupMembers :: Lens' DeregisterTransitGatewayMulticastGroupMembersResponse (Maybe TransitGatewayMulticastDeregisteredGroupMembers)

Information about the deregistered members.

deregisterTransitGatewayMulticastGroupMembersResponse_httpStatus :: Lens' DeregisterTransitGatewayMulticastGroupMembersResponse Int

The response's http status code.