Network.AWS.EC2.DeregisterTransitGatewayMulticastGroupSources

Description

Deregisters the specified sources (network interfaces) from the transit gateway multicast group.

Creating a Request

data DeregisterTransitGatewayMulticastGroupSources

See: newDeregisterTransitGatewayMulticastGroupSources smart constructor.

Constructors

DeregisterTransitGatewayMulticastGroupSources'

Fields

transitGatewayMulticastDomainId :: Maybe Text

The ID of the transit gateway multicast domain.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

networkInterfaceIds :: Maybe [Text]

The IDs of the group sources' network interfaces.

groupIpAddress :: Maybe Text

The IP address assigned to the transit gateway multicast group.

newDeregisterTransitGatewayMulticastGroupSources :: DeregisterTransitGatewayMulticastGroupSources

Create a value of DeregisterTransitGatewayMulticastGroupSources with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:transitGatewayMulticastDomainId:DeregisterTransitGatewayMulticastGroupSources', deregisterTransitGatewayMulticastGroupSources_transitGatewayMulticastDomainId - The ID of the transit gateway multicast domain.

$sel:dryRun:DeregisterTransitGatewayMulticastGroupSources', deregisterTransitGatewayMulticastGroupSources_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:networkInterfaceIds:DeregisterTransitGatewayMulticastGroupSources', deregisterTransitGatewayMulticastGroupSources_networkInterfaceIds - The IDs of the group sources' network interfaces.

$sel:groupIpAddress:DeregisterTransitGatewayMulticastGroupSources', deregisterTransitGatewayMulticastGroupSources_groupIpAddress - The IP address assigned to the transit gateway multicast group.

Request Lenses

deregisterTransitGatewayMulticastGroupSources_transitGatewayMulticastDomainId :: Lens' DeregisterTransitGatewayMulticastGroupSources (Maybe Text)

The ID of the transit gateway multicast domain.

deregisterTransitGatewayMulticastGroupSources_dryRun :: Lens' DeregisterTransitGatewayMulticastGroupSources (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

deregisterTransitGatewayMulticastGroupSources_networkInterfaceIds :: Lens' DeregisterTransitGatewayMulticastGroupSources (Maybe [Text])

The IDs of the group sources' network interfaces.

deregisterTransitGatewayMulticastGroupSources_groupIpAddress :: Lens' DeregisterTransitGatewayMulticastGroupSources (Maybe Text)

The IP address assigned to the transit gateway multicast group.

Destructuring the Response

data DeregisterTransitGatewayMulticastGroupSourcesResponse

See: newDeregisterTransitGatewayMulticastGroupSourcesResponse smart constructor.

Constructors

DeregisterTransitGatewayMulticastGroupSourcesResponse'

Fields

deregisteredMulticastGroupSources :: Maybe TransitGatewayMulticastDeregisteredGroupSources

Information about the deregistered group sources.

httpStatus :: Int

The response's http status code.

newDeregisterTransitGatewayMulticastGroupSourcesResponse

Arguments

:: Int
$sel:httpStatus:DeregisterTransitGatewayMulticastGroupSourcesResponse'

-> DeregisterTransitGatewayMulticastGroupSourcesResponse

Create a value of DeregisterTransitGatewayMulticastGroupSourcesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deregisteredMulticastGroupSources:DeregisterTransitGatewayMulticastGroupSourcesResponse', deregisterTransitGatewayMulticastGroupSourcesResponse_deregisteredMulticastGroupSources - Information about the deregistered group sources.

$sel:httpStatus:DeregisterTransitGatewayMulticastGroupSourcesResponse', deregisterTransitGatewayMulticastGroupSourcesResponse_httpStatus - The response's http status code.

Response Lenses

deregisterTransitGatewayMulticastGroupSourcesResponse_deregisteredMulticastGroupSources :: Lens' DeregisterTransitGatewayMulticastGroupSourcesResponse (Maybe TransitGatewayMulticastDeregisteredGroupSources)

Information about the deregistered group sources.

deregisterTransitGatewayMulticastGroupSourcesResponse_httpStatus :: Lens' DeregisterTransitGatewayMulticastGroupSourcesResponse Int

The response's http status code.