Network.AWS.EC2.DescribeAccountAttributes

Description

Describes attributes of your AWS account. The following are the supported account attributes:

supported-platforms: Indicates whether your account can launch instances into EC2-Classic and EC2-VPC, or only into EC2-VPC.

default-vpc: The ID of the default VPC for your account, or none.

max-instances: This attribute is no longer supported. The returned value does not reflect your actual vCPU limit for running On-Demand Instances. For more information, see On-Demand Instance Limits in the Amazon Elastic Compute Cloud User Guide.

vpc-max-security-groups-per-interface: The maximum number of security groups that you can assign to a network interface.

max-elastic-ips: The maximum number of Elastic IP addresses that you can allocate for use with EC2-Classic.

vpc-max-elastic-ips: The maximum number of Elastic IP addresses that you can allocate for use with EC2-VPC.

Creating a Request

data DescribeAccountAttributes

See: newDescribeAccountAttributes smart constructor.

Constructors

DescribeAccountAttributes'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

attributeNames :: Maybe [AccountAttributeName]

The account attribute names.

newDescribeAccountAttributes :: DescribeAccountAttributes

Create a value of DescribeAccountAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:DescribeAccountAttributes', describeAccountAttributes_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:attributeNames:DescribeAccountAttributes', describeAccountAttributes_attributeNames - The account attribute names.

Request Lenses

describeAccountAttributes_dryRun :: Lens' DescribeAccountAttributes (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeAccountAttributes_attributeNames :: Lens' DescribeAccountAttributes (Maybe [AccountAttributeName])

The account attribute names.

Destructuring the Response

data DescribeAccountAttributesResponse

See: newDescribeAccountAttributesResponse smart constructor.

Constructors

DescribeAccountAttributesResponse'

Fields

accountAttributes :: Maybe [AccountAttribute]

Information about the account attributes.

httpStatus :: Int

The response's http status code.

newDescribeAccountAttributesResponse

Arguments

:: Int
$sel:httpStatus:DescribeAccountAttributesResponse'

-> DescribeAccountAttributesResponse

Create a value of DescribeAccountAttributesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountAttributes:DescribeAccountAttributesResponse', describeAccountAttributesResponse_accountAttributes - Information about the account attributes.

$sel:httpStatus:DescribeAccountAttributesResponse', describeAccountAttributesResponse_httpStatus - The response's http status code.

Response Lenses

describeAccountAttributesResponse_accountAttributes :: Lens' DescribeAccountAttributesResponse (Maybe [AccountAttribute])

Information about the account attributes.

describeAccountAttributesResponse_httpStatus :: Lens' DescribeAccountAttributesResponse Int

The response's http status code.