Network.AWS.EC2.DescribeAddresses

Description

Describes the specified Elastic IP addresses or all of your Elastic IP addresses.

An Elastic IP address is for use in either the EC2-Classic platform or in a VPC. For more information, see Elastic IP Addresses in the Amazon Elastic Compute Cloud User Guide.

Creating a Request

data DescribeAddresses

See: newDescribeAddresses smart constructor.

Constructors

DescribeAddresses'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

allocationIds :: Maybe [Text]

EC2-VPC

Information about the allocation IDs.

publicIps :: Maybe [Text]

One or more Elastic IP addresses.

Default: Describes all your Elastic IP addresses.

filters :: Maybe [Filter]

One or more filters. Filter names and values are case-sensitive.

allocation-id - [EC2-VPC] The allocation ID for the address.

association-id - [EC2-VPC] The association ID for the address.

domain - Indicates whether the address is for use in EC2-Classic (standard) or in a VPC (vpc).

instance-id - The ID of the instance the address is associated with, if any.

network-border-group - A unique set of Availability Zones, Local Zones, or Wavelength Zones from where Amazon Web Services advertises IP addresses.

network-interface-id - [EC2-VPC] The ID of the network interface that the address is associated with, if any.

network-interface-owner-id - The Amazon Web Services account ID of the owner.

private-ip-address - [EC2-VPC] The private IP address associated with the Elastic IP address.

public-ip - The Elastic IP address, or the carrier IP address.

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

newDescribeAddresses :: DescribeAddresses

Create a value of DescribeAddresses with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:DescribeAddresses', describeAddresses_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:allocationIds:DescribeAddresses', describeAddresses_allocationIds - [EC2-VPC] Information about the allocation IDs.

$sel:publicIps:DescribeAddresses', describeAddresses_publicIps - One or more Elastic IP addresses.

Default: Describes all your Elastic IP addresses.

$sel:filters:DescribeAddresses', describeAddresses_filters - One or more filters. Filter names and values are case-sensitive.

allocation-id - [EC2-VPC] The allocation ID for the address.

association-id - [EC2-VPC] The association ID for the address.

domain - Indicates whether the address is for use in EC2-Classic (standard) or in a VPC (vpc).

instance-id - The ID of the instance the address is associated with, if any.

network-border-group - A unique set of Availability Zones, Local Zones, or Wavelength Zones from where Amazon Web Services advertises IP addresses.

network-interface-id - [EC2-VPC] The ID of the network interface that the address is associated with, if any.

network-interface-owner-id - The Amazon Web Services account ID of the owner.

private-ip-address - [EC2-VPC] The private IP address associated with the Elastic IP address.

public-ip - The Elastic IP address, or the carrier IP address.

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

Request Lenses

describeAddresses_dryRun :: Lens' DescribeAddresses (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeAddresses_allocationIds :: Lens' DescribeAddresses (Maybe [Text])

EC2-VPC

Information about the allocation IDs.

describeAddresses_publicIps :: Lens' DescribeAddresses (Maybe [Text])

One or more Elastic IP addresses.

Default: Describes all your Elastic IP addresses.

describeAddresses_filters :: Lens' DescribeAddresses (Maybe [Filter])

One or more filters. Filter names and values are case-sensitive.

allocation-id - [EC2-VPC] The allocation ID for the address.

association-id - [EC2-VPC] The association ID for the address.

domain - Indicates whether the address is for use in EC2-Classic (standard) or in a VPC (vpc).

instance-id - The ID of the instance the address is associated with, if any.

network-border-group - A unique set of Availability Zones, Local Zones, or Wavelength Zones from where Amazon Web Services advertises IP addresses.

network-interface-id - [EC2-VPC] The ID of the network interface that the address is associated with, if any.

network-interface-owner-id - The Amazon Web Services account ID of the owner.

private-ip-address - [EC2-VPC] The private IP address associated with the Elastic IP address.

public-ip - The Elastic IP address, or the carrier IP address.

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

Destructuring the Response

data DescribeAddressesResponse

See: newDescribeAddressesResponse smart constructor.

Constructors

DescribeAddressesResponse'

Fields

addresses :: Maybe [Address]

Information about the Elastic IP addresses.

httpStatus :: Int

The response's http status code.

newDescribeAddressesResponse

Arguments

:: Int
$sel:httpStatus:DescribeAddressesResponse'

-> DescribeAddressesResponse

Create a value of DescribeAddressesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:addresses:DescribeAddressesResponse', describeAddressesResponse_addresses - Information about the Elastic IP addresses.

$sel:httpStatus:DescribeAddressesResponse', describeAddressesResponse_httpStatus - The response's http status code.

Response Lenses

describeAddressesResponse_addresses :: Lens' DescribeAddressesResponse (Maybe [Address])

Information about the Elastic IP addresses.

describeAddressesResponse_httpStatus :: Lens' DescribeAddressesResponse Int

The response's http status code.