Network.AWS.EC2.DescribeAddressesAttribute

Description

Describes the attributes of the specified Elastic IP addresses. For requirements, see Using reverse DNS for email applications.

This operation returns paginated results.

Creating a Request

data DescribeAddressesAttribute

See: newDescribeAddressesAttribute smart constructor.

Constructors

DescribeAddressesAttribute'

Fields

attribute :: Maybe AddressAttributeName

The attribute of the IP address.

nextToken :: Maybe Text

The token for the next page of results.

allocationIds :: Maybe [Text]

EC2-VPC

The allocation IDs.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

maxResults :: Maybe Natural

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

newDescribeAddressesAttribute :: DescribeAddressesAttribute

Create a value of DescribeAddressesAttribute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attribute:DescribeAddressesAttribute', describeAddressesAttribute_attribute - The attribute of the IP address.

$sel:nextToken:DescribeAddressesAttribute', describeAddressesAttribute_nextToken - The token for the next page of results.

$sel:allocationIds:DescribeAddressesAttribute', describeAddressesAttribute_allocationIds - [EC2-VPC] The allocation IDs.

$sel:dryRun:DescribeAddressesAttribute', describeAddressesAttribute_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:maxResults:DescribeAddressesAttribute', describeAddressesAttribute_maxResults - The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

Request Lenses

describeAddressesAttribute_attribute :: Lens' DescribeAddressesAttribute (Maybe AddressAttributeName)

The attribute of the IP address.

describeAddressesAttribute_nextToken :: Lens' DescribeAddressesAttribute (Maybe Text)

The token for the next page of results.

describeAddressesAttribute_allocationIds :: Lens' DescribeAddressesAttribute (Maybe [Text])

EC2-VPC

The allocation IDs.

describeAddressesAttribute_dryRun :: Lens' DescribeAddressesAttribute (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeAddressesAttribute_maxResults :: Lens' DescribeAddressesAttribute (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

Destructuring the Response

data DescribeAddressesAttributeResponse

See: newDescribeAddressesAttributeResponse smart constructor.

Constructors

DescribeAddressesAttributeResponse'

Fields

addresses :: Maybe [AddressAttribute]

Information about the IP addresses.

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

httpStatus :: Int

The response's http status code.

newDescribeAddressesAttributeResponse

Arguments

:: Int
$sel:httpStatus:DescribeAddressesAttributeResponse'

-> DescribeAddressesAttributeResponse

Create a value of DescribeAddressesAttributeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:addresses:DescribeAddressesAttributeResponse', describeAddressesAttributeResponse_addresses - Information about the IP addresses.

$sel:nextToken:DescribeAddressesAttribute', describeAddressesAttributeResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:httpStatus:DescribeAddressesAttributeResponse', describeAddressesAttributeResponse_httpStatus - The response's http status code.

Response Lenses

describeAddressesAttributeResponse_addresses :: Lens' DescribeAddressesAttributeResponse (Maybe [AddressAttribute])

Information about the IP addresses.

describeAddressesAttributeResponse_nextToken :: Lens' DescribeAddressesAttributeResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

describeAddressesAttributeResponse_httpStatus :: Lens' DescribeAddressesAttributeResponse Int

The response's http status code.