Network.AWS.EC2.DescribeBundleTasks

Description

Describes the specified bundle tasks or all of your bundle tasks.

Completed bundle tasks are listed for only a limited time. If your bundle task is no longer in the list, you can still register an AMI from it. Just use RegisterImage with the Amazon S3 bucket name and image manifest name you provided to the bundle task.

Creating a Request

data DescribeBundleTasks

See: newDescribeBundleTasks smart constructor.

Constructors

DescribeBundleTasks'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

filters :: Maybe [Filter]

The filters.

bundle-id - The ID of the bundle task.

error-code - If the task failed, the error code returned.

error-message - If the task failed, the error message returned.

instance-id - The ID of the instance.

progress - The level of task completion, as a percentage (for example, 20%).

s3-bucket - The Amazon S3 bucket to store the AMI.

s3-prefix - The beginning of the AMI name.

start-time - The time the task started (for example, 2013-09-15T17:15:20.000Z).

state - The state of the task (pending | waiting-for-shutdown | bundling | storing | cancelling | complete | failed).

update-time - The time of the most recent update for the task.

bundleIds :: Maybe [Text]

The bundle task IDs.

Default: Describes all your bundle tasks.

newDescribeBundleTasks :: DescribeBundleTasks

Create a value of DescribeBundleTasks with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:DescribeBundleTasks', describeBundleTasks_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:filters:DescribeBundleTasks', describeBundleTasks_filters - The filters.

bundle-id - The ID of the bundle task.

error-code - If the task failed, the error code returned.

error-message - If the task failed, the error message returned.

instance-id - The ID of the instance.

progress - The level of task completion, as a percentage (for example, 20%).

s3-bucket - The Amazon S3 bucket to store the AMI.

s3-prefix - The beginning of the AMI name.

start-time - The time the task started (for example, 2013-09-15T17:15:20.000Z).

state - The state of the task (pending | waiting-for-shutdown | bundling | storing | cancelling | complete | failed).

update-time - The time of the most recent update for the task.

$sel:bundleIds:DescribeBundleTasks', describeBundleTasks_bundleIds - The bundle task IDs.

Default: Describes all your bundle tasks.

Request Lenses

describeBundleTasks_dryRun :: Lens' DescribeBundleTasks (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeBundleTasks_filters :: Lens' DescribeBundleTasks (Maybe [Filter])

The filters.

bundle-id - The ID of the bundle task.

error-code - If the task failed, the error code returned.

error-message - If the task failed, the error message returned.

instance-id - The ID of the instance.

progress - The level of task completion, as a percentage (for example, 20%).

s3-bucket - The Amazon S3 bucket to store the AMI.

s3-prefix - The beginning of the AMI name.

start-time - The time the task started (for example, 2013-09-15T17:15:20.000Z).

state - The state of the task (pending | waiting-for-shutdown | bundling | storing | cancelling | complete | failed).

update-time - The time of the most recent update for the task.

describeBundleTasks_bundleIds :: Lens' DescribeBundleTasks (Maybe [Text])

The bundle task IDs.

Default: Describes all your bundle tasks.

Destructuring the Response

data DescribeBundleTasksResponse

See: newDescribeBundleTasksResponse smart constructor.

Constructors

DescribeBundleTasksResponse'

Fields

bundleTasks :: Maybe [BundleTask]

Information about the bundle tasks.

httpStatus :: Int

The response's http status code.

newDescribeBundleTasksResponse

Arguments

:: Int
$sel:httpStatus:DescribeBundleTasksResponse'

-> DescribeBundleTasksResponse

Create a value of DescribeBundleTasksResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bundleTasks:DescribeBundleTasksResponse', describeBundleTasksResponse_bundleTasks - Information about the bundle tasks.

$sel:httpStatus:DescribeBundleTasksResponse', describeBundleTasksResponse_httpStatus - The response's http status code.

Response Lenses

describeBundleTasksResponse_bundleTasks :: Lens' DescribeBundleTasksResponse (Maybe [BundleTask])

Information about the bundle tasks.

describeBundleTasksResponse_httpStatus :: Lens' DescribeBundleTasksResponse Int

The response's http status code.