Network.AWS.EC2.DescribeDhcpOptions

Description

Describes one or more of your DHCP options sets.

For more information, see DHCP options sets in the Amazon Virtual Private Cloud User Guide.

This operation returns paginated results.

Creating a Request

data DescribeDhcpOptions

See: newDescribeDhcpOptions smart constructor.

Constructors

DescribeDhcpOptions'

Fields

filters :: Maybe [Filter]

One or more filters.

dhcp-options-id - The ID of a DHCP options set.

key - The key for one of the options (for example, domain-name).

value - The value for one of the options.

owner-id - The ID of the Amazon Web Services account that owns the DHCP options set.

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

dhcpOptionsIds :: Maybe [Text]

The IDs of one or more DHCP options sets.

Default: Describes all your DHCP options sets.

nextToken :: Maybe Text

The token for the next page of results.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

maxResults :: Maybe Natural

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

newDescribeDhcpOptions :: DescribeDhcpOptions

Create a value of DescribeDhcpOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:DescribeDhcpOptions', describeDhcpOptions_filters - One or more filters.

dhcp-options-id - The ID of a DHCP options set.

key - The key for one of the options (for example, domain-name).

value - The value for one of the options.

owner-id - The ID of the Amazon Web Services account that owns the DHCP options set.

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

$sel:dhcpOptionsIds:DescribeDhcpOptions', describeDhcpOptions_dhcpOptionsIds - The IDs of one or more DHCP options sets.

Default: Describes all your DHCP options sets.

$sel:nextToken:DescribeDhcpOptions', describeDhcpOptions_nextToken - The token for the next page of results.

$sel:dryRun:DescribeDhcpOptions', describeDhcpOptions_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:maxResults:DescribeDhcpOptions', describeDhcpOptions_maxResults - The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

Request Lenses

describeDhcpOptions_filters :: Lens' DescribeDhcpOptions (Maybe [Filter])

One or more filters.

dhcp-options-id - The ID of a DHCP options set.

key - The key for one of the options (for example, domain-name).

value - The value for one of the options.

owner-id - The ID of the Amazon Web Services account that owns the DHCP options set.

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

describeDhcpOptions_dhcpOptionsIds :: Lens' DescribeDhcpOptions (Maybe [Text])

The IDs of one or more DHCP options sets.

Default: Describes all your DHCP options sets.

describeDhcpOptions_nextToken :: Lens' DescribeDhcpOptions (Maybe Text)

The token for the next page of results.

describeDhcpOptions_dryRun :: Lens' DescribeDhcpOptions (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeDhcpOptions_maxResults :: Lens' DescribeDhcpOptions (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

Destructuring the Response

data DescribeDhcpOptionsResponse

See: newDescribeDhcpOptionsResponse smart constructor.

Constructors

DescribeDhcpOptionsResponse'

Fields

dhcpOptions :: Maybe [DhcpOptions]

Information about one or more DHCP options sets.

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

httpStatus :: Int

The response's http status code.

newDescribeDhcpOptionsResponse

Arguments

:: Int
$sel:httpStatus:DescribeDhcpOptionsResponse'

-> DescribeDhcpOptionsResponse

Create a value of DescribeDhcpOptionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dhcpOptions:DescribeDhcpOptionsResponse', describeDhcpOptionsResponse_dhcpOptions - Information about one or more DHCP options sets.

$sel:nextToken:DescribeDhcpOptions', describeDhcpOptionsResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:httpStatus:DescribeDhcpOptionsResponse', describeDhcpOptionsResponse_httpStatus - The response's http status code.

Response Lenses

describeDhcpOptionsResponse_dhcpOptions :: Lens' DescribeDhcpOptionsResponse (Maybe [DhcpOptions])

Information about one or more DHCP options sets.

describeDhcpOptionsResponse_nextToken :: Lens' DescribeDhcpOptionsResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

describeDhcpOptionsResponse_httpStatus :: Lens' DescribeDhcpOptionsResponse Int

The response's http status code.