Network.AWS.EC2.DescribeFleetHistory

Description

Describes the events for the specified EC2 Fleet during the specified time.

EC2 Fleet events are delayed by up to 30 seconds before they can be described. This ensures that you can query by the last evaluated time and not miss a recorded event. EC2 Fleet events are available for 48 hours.

For more information, see Monitoring your EC2 Fleet in the Amazon EC2 User Guide.

Creating a Request

data DescribeFleetHistory

See: newDescribeFleetHistory smart constructor.

Constructors

DescribeFleetHistory'

Fields

eventType :: Maybe FleetEventType

The type of events to describe. By default, all events are described.

nextToken :: Maybe Text

The token for the next set of results.

maxResults :: Maybe Int

The maximum number of results to return in a single call. Specify a value between 1 and 1000. The default value is 1000. To retrieve the remaining results, make another call with the returned NextToken value.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

fleetId :: Text

The ID of the EC2 Fleet.

startTime :: ISO8601

The start date and time for the events, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ).

newDescribeFleetHistory

Arguments

:: Text
$sel:fleetId:DescribeFleetHistory'

-> UTCTime
$sel:startTime:DescribeFleetHistory'

-> DescribeFleetHistory

Create a value of DescribeFleetHistory with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventType:DescribeFleetHistory', describeFleetHistory_eventType - The type of events to describe. By default, all events are described.

$sel:nextToken:DescribeFleetHistory', describeFleetHistory_nextToken - The token for the next set of results.

$sel:maxResults:DescribeFleetHistory', describeFleetHistory_maxResults - The maximum number of results to return in a single call. Specify a value between 1 and 1000. The default value is 1000. To retrieve the remaining results, make another call with the returned NextToken value.

$sel:dryRun:DescribeFleetHistory', describeFleetHistory_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:fleetId:DescribeFleetHistory', describeFleetHistory_fleetId - The ID of the EC2 Fleet.

$sel:startTime:DescribeFleetHistory', describeFleetHistory_startTime - The start date and time for the events, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ).

Request Lenses

describeFleetHistory_eventType :: Lens' DescribeFleetHistory (Maybe FleetEventType)

The type of events to describe. By default, all events are described.

describeFleetHistory_nextToken :: Lens' DescribeFleetHistory (Maybe Text)

The token for the next set of results.

describeFleetHistory_maxResults :: Lens' DescribeFleetHistory (Maybe Int)

The maximum number of results to return in a single call. Specify a value between 1 and 1000. The default value is 1000. To retrieve the remaining results, make another call with the returned NextToken value.

describeFleetHistory_dryRun :: Lens' DescribeFleetHistory (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeFleetHistory_fleetId :: Lens' DescribeFleetHistory Text

The ID of the EC2 Fleet.

describeFleetHistory_startTime :: Lens' DescribeFleetHistory UTCTime

The start date and time for the events, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ).

Destructuring the Response

data DescribeFleetHistoryResponse

See: newDescribeFleetHistoryResponse smart constructor.

Constructors

DescribeFleetHistoryResponse'

Fields

nextToken :: Maybe Text

The token for the next set of results.

fleetId :: Maybe Text

The ID of the EC Fleet.

startTime :: Maybe ISO8601

The start date and time for the events, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ).

historyRecords :: Maybe [HistoryRecordEntry]

Information about the events in the history of the EC2 Fleet.

lastEvaluatedTime :: Maybe ISO8601

The last date and time for the events, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ). All records up to this time were retrieved.

If nextToken indicates that there are more results, this value is not present.

httpStatus :: Int

The response's http status code.

newDescribeFleetHistoryResponse

Arguments

:: Int
$sel:httpStatus:DescribeFleetHistoryResponse'

-> DescribeFleetHistoryResponse

Create a value of DescribeFleetHistoryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeFleetHistory', describeFleetHistoryResponse_nextToken - The token for the next set of results.

$sel:fleetId:DescribeFleetHistory', describeFleetHistoryResponse_fleetId - The ID of the EC Fleet.

$sel:startTime:DescribeFleetHistory', describeFleetHistoryResponse_startTime - The start date and time for the events, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ).

$sel:historyRecords:DescribeFleetHistoryResponse', describeFleetHistoryResponse_historyRecords - Information about the events in the history of the EC2 Fleet.

$sel:lastEvaluatedTime:DescribeFleetHistoryResponse', describeFleetHistoryResponse_lastEvaluatedTime - The last date and time for the events, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ). All records up to this time were retrieved.

If nextToken indicates that there are more results, this value is not present.

$sel:httpStatus:DescribeFleetHistoryResponse', describeFleetHistoryResponse_httpStatus - The response's http status code.

Response Lenses

describeFleetHistoryResponse_nextToken :: Lens' DescribeFleetHistoryResponse (Maybe Text)

The token for the next set of results.

describeFleetHistoryResponse_fleetId :: Lens' DescribeFleetHistoryResponse (Maybe Text)

The ID of the EC Fleet.

describeFleetHistoryResponse_startTime :: Lens' DescribeFleetHistoryResponse (Maybe UTCTime)

The start date and time for the events, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ).

describeFleetHistoryResponse_historyRecords :: Lens' DescribeFleetHistoryResponse (Maybe [HistoryRecordEntry])

Information about the events in the history of the EC2 Fleet.

describeFleetHistoryResponse_lastEvaluatedTime :: Lens' DescribeFleetHistoryResponse (Maybe UTCTime)

The last date and time for the events, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ). All records up to this time were retrieved.

If nextToken indicates that there are more results, this value is not present.

describeFleetHistoryResponse_httpStatus :: Lens' DescribeFleetHistoryResponse Int

The response's http status code.