Network.AWS.EC2.DescribeFleetInstances

Description

Describes the running instances for the specified EC2 Fleet.

For more information, see Monitoring your EC2 Fleet in the Amazon EC2 User Guide.

Creating a Request

data DescribeFleetInstances

See: newDescribeFleetInstances smart constructor.

Constructors

DescribeFleetInstances'

Fields

filters :: Maybe [Filter]

The filters.

instance-type - The instance type.

nextToken :: Maybe Text

The token for the next set of results.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

maxResults :: Maybe Int

The maximum number of results to return in a single call. Specify a value between 1 and 1000. The default value is 1000. To retrieve the remaining results, make another call with the returned NextToken value.

fleetId :: Text

The ID of the EC2 Fleet.

newDescribeFleetInstances

Arguments

:: Text
$sel:fleetId:DescribeFleetInstances'

-> DescribeFleetInstances

Create a value of DescribeFleetInstances with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:DescribeFleetInstances', describeFleetInstances_filters - The filters.

instance-type - The instance type.

$sel:nextToken:DescribeFleetInstances', describeFleetInstances_nextToken - The token for the next set of results.

$sel:dryRun:DescribeFleetInstances', describeFleetInstances_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:maxResults:DescribeFleetInstances', describeFleetInstances_maxResults - The maximum number of results to return in a single call. Specify a value between 1 and 1000. The default value is 1000. To retrieve the remaining results, make another call with the returned NextToken value.

$sel:fleetId:DescribeFleetInstances', describeFleetInstances_fleetId - The ID of the EC2 Fleet.

Request Lenses

describeFleetInstances_filters :: Lens' DescribeFleetInstances (Maybe [Filter])

The filters.

instance-type - The instance type.

describeFleetInstances_nextToken :: Lens' DescribeFleetInstances (Maybe Text)

The token for the next set of results.

describeFleetInstances_dryRun :: Lens' DescribeFleetInstances (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeFleetInstances_maxResults :: Lens' DescribeFleetInstances (Maybe Int)

The maximum number of results to return in a single call. Specify a value between 1 and 1000. The default value is 1000. To retrieve the remaining results, make another call with the returned NextToken value.

describeFleetInstances_fleetId :: Lens' DescribeFleetInstances Text

The ID of the EC2 Fleet.

Destructuring the Response

data DescribeFleetInstancesResponse

See: newDescribeFleetInstancesResponse smart constructor.

Constructors

DescribeFleetInstancesResponse'

Fields

nextToken :: Maybe Text

The token for the next set of results.

fleetId :: Maybe Text

The ID of the EC2 Fleet.

activeInstances :: Maybe [ActiveInstance]

The running instances. This list is refreshed periodically and might be out of date.

httpStatus :: Int

The response's http status code.

newDescribeFleetInstancesResponse

Arguments

:: Int
$sel:httpStatus:DescribeFleetInstancesResponse'

-> DescribeFleetInstancesResponse

Create a value of DescribeFleetInstancesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeFleetInstances', describeFleetInstancesResponse_nextToken - The token for the next set of results.

$sel:fleetId:DescribeFleetInstances', describeFleetInstancesResponse_fleetId - The ID of the EC2 Fleet.

$sel:activeInstances:DescribeFleetInstancesResponse', describeFleetInstancesResponse_activeInstances - The running instances. This list is refreshed periodically and might be out of date.

$sel:httpStatus:DescribeFleetInstancesResponse', describeFleetInstancesResponse_httpStatus - The response's http status code.

Response Lenses

describeFleetInstancesResponse_nextToken :: Lens' DescribeFleetInstancesResponse (Maybe Text)

The token for the next set of results.

describeFleetInstancesResponse_fleetId :: Lens' DescribeFleetInstancesResponse (Maybe Text)

The ID of the EC2 Fleet.

describeFleetInstancesResponse_activeInstances :: Lens' DescribeFleetInstancesResponse (Maybe [ActiveInstance])

The running instances. This list is refreshed periodically and might be out of date.

describeFleetInstancesResponse_httpStatus :: Lens' DescribeFleetInstancesResponse Int

The response's http status code.