Network.AWS.EC2.DescribeFleets

Description

Describes the specified EC2 Fleets or all of your EC2 Fleets.

For more information, see Monitoring your EC2 Fleet in the Amazon EC2 User Guide.

This operation returns paginated results.

Creating a Request

data DescribeFleets

See: newDescribeFleets smart constructor.

Constructors

DescribeFleets'

Fields

nextToken :: Maybe Text

The token for the next set of results.

maxResults :: Maybe Int

The maximum number of results to return in a single call. Specify a value between 1 and 1000. The default value is 1000. To retrieve the remaining results, make another call with the returned NextToken value.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

filters :: Maybe [Filter]

The filters.

activity-status - The progress of the EC2 Fleet ( error | pending-fulfillment | pending-termination | fulfilled).

excess-capacity-termination-policy - Indicates whether to terminate running instances if the target capacity is decreased below the current EC2 Fleet size (true | false).

fleet-state - The state of the EC2 Fleet (submitted | active | deleted | failed | deleted-running | deleted-terminating | modifying).

replace-unhealthy-instances - Indicates whether EC2 Fleet should replace unhealthy instances (true | false).

type - The type of request (instant | request | maintain).

fleetIds :: Maybe [Text]

The ID of the EC2 Fleets.

newDescribeFleets :: DescribeFleets

Create a value of DescribeFleets with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeFleets', describeFleets_nextToken - The token for the next set of results.

$sel:maxResults:DescribeFleets', describeFleets_maxResults - The maximum number of results to return in a single call. Specify a value between 1 and 1000. The default value is 1000. To retrieve the remaining results, make another call with the returned NextToken value.

$sel:dryRun:DescribeFleets', describeFleets_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:filters:DescribeFleets', describeFleets_filters - The filters.

activity-status - The progress of the EC2 Fleet ( error | pending-fulfillment | pending-termination | fulfilled).

excess-capacity-termination-policy - Indicates whether to terminate running instances if the target capacity is decreased below the current EC2 Fleet size (true | false).

fleet-state - The state of the EC2 Fleet (submitted | active | deleted | failed | deleted-running | deleted-terminating | modifying).

replace-unhealthy-instances - Indicates whether EC2 Fleet should replace unhealthy instances (true | false).

type - The type of request (instant | request | maintain).

$sel:fleetIds:DescribeFleets', describeFleets_fleetIds - The ID of the EC2 Fleets.

Request Lenses

describeFleets_nextToken :: Lens' DescribeFleets (Maybe Text)

The token for the next set of results.

describeFleets_maxResults :: Lens' DescribeFleets (Maybe Int)

The maximum number of results to return in a single call. Specify a value between 1 and 1000. The default value is 1000. To retrieve the remaining results, make another call with the returned NextToken value.

describeFleets_dryRun :: Lens' DescribeFleets (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeFleets_filters :: Lens' DescribeFleets (Maybe [Filter])

The filters.

activity-status - The progress of the EC2 Fleet ( error | pending-fulfillment | pending-termination | fulfilled).

excess-capacity-termination-policy - Indicates whether to terminate running instances if the target capacity is decreased below the current EC2 Fleet size (true | false).

fleet-state - The state of the EC2 Fleet (submitted | active | deleted | failed | deleted-running | deleted-terminating | modifying).

replace-unhealthy-instances - Indicates whether EC2 Fleet should replace unhealthy instances (true | false).

type - The type of request (instant | request | maintain).

describeFleets_fleetIds :: Lens' DescribeFleets (Maybe [Text])

The ID of the EC2 Fleets.

Destructuring the Response

data DescribeFleetsResponse

See: newDescribeFleetsResponse smart constructor.

Constructors

DescribeFleetsResponse'

Fields

nextToken :: Maybe Text

The token for the next set of results.

fleets :: Maybe [FleetData]

Information about the EC2 Fleets.

httpStatus :: Int

The response's http status code.

newDescribeFleetsResponse

Arguments

:: Int
$sel:httpStatus:DescribeFleetsResponse'

-> DescribeFleetsResponse

Create a value of DescribeFleetsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeFleets', describeFleetsResponse_nextToken - The token for the next set of results.

$sel:fleets:DescribeFleetsResponse', describeFleetsResponse_fleets - Information about the EC2 Fleets.

$sel:httpStatus:DescribeFleetsResponse', describeFleetsResponse_httpStatus - The response's http status code.

Response Lenses

describeFleetsResponse_nextToken :: Lens' DescribeFleetsResponse (Maybe Text)

The token for the next set of results.

describeFleetsResponse_fleets :: Lens' DescribeFleetsResponse (Maybe [FleetData])

Information about the EC2 Fleets.

describeFleetsResponse_httpStatus :: Lens' DescribeFleetsResponse Int

The response's http status code.