Network.AWS.EC2.DescribeHostReservations

Description

Describes reservations that are associated with Dedicated Hosts in your account.

This operation returns paginated results.

Creating a Request

data DescribeHostReservations

See: newDescribeHostReservations smart constructor.

Constructors

DescribeHostReservations'

Fields

nextToken :: Maybe Text

The token to use to retrieve the next page of results.

hostReservationIdSet :: Maybe [Text]

The host reservation IDs.

filter' :: Maybe [Filter]

The filters.

instance-family - The instance family (for example, m4).

payment-option - The payment option (NoUpfront | PartialUpfront | AllUpfront).

state - The state of the reservation (payment-pending | payment-failed | active | retired).

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

maxResults :: Maybe Int

The maximum number of results to return for the request in a single page. The remaining results can be seen by sending another request with the returned nextToken value. This value can be between 5 and 500. If maxResults is given a larger value than 500, you receive an error.

newDescribeHostReservations :: DescribeHostReservations

Create a value of DescribeHostReservations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeHostReservations', describeHostReservations_nextToken - The token to use to retrieve the next page of results.

$sel:hostReservationIdSet:DescribeHostReservations', describeHostReservations_hostReservationIdSet - The host reservation IDs.

$sel:filter':DescribeHostReservations', describeHostReservations_filter - The filters.

instance-family - The instance family (for example, m4).

payment-option - The payment option (NoUpfront | PartialUpfront | AllUpfront).

state - The state of the reservation (payment-pending | payment-failed | active | retired).

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

$sel:maxResults:DescribeHostReservations', describeHostReservations_maxResults - The maximum number of results to return for the request in a single page. The remaining results can be seen by sending another request with the returned nextToken value. This value can be between 5 and 500. If maxResults is given a larger value than 500, you receive an error.

Request Lenses

describeHostReservations_nextToken :: Lens' DescribeHostReservations (Maybe Text)

The token to use to retrieve the next page of results.

describeHostReservations_hostReservationIdSet :: Lens' DescribeHostReservations (Maybe [Text])

The host reservation IDs.

describeHostReservations_filter :: Lens' DescribeHostReservations (Maybe [Filter])

The filters.

instance-family - The instance family (for example, m4).

payment-option - The payment option (NoUpfront | PartialUpfront | AllUpfront).

state - The state of the reservation (payment-pending | payment-failed | active | retired).

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

describeHostReservations_maxResults :: Lens' DescribeHostReservations (Maybe Int)

The maximum number of results to return for the request in a single page. The remaining results can be seen by sending another request with the returned nextToken value. This value can be between 5 and 500. If maxResults is given a larger value than 500, you receive an error.

Destructuring the Response

data DescribeHostReservationsResponse

See: newDescribeHostReservationsResponse smart constructor.

Constructors

DescribeHostReservationsResponse'

Fields

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

hostReservationSet :: Maybe [HostReservation]

Details about the reservation's configuration.

httpStatus :: Int

The response's http status code.

newDescribeHostReservationsResponse

Arguments

:: Int
$sel:httpStatus:DescribeHostReservationsResponse'

-> DescribeHostReservationsResponse

Create a value of DescribeHostReservationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeHostReservations', describeHostReservationsResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:hostReservationSet:DescribeHostReservationsResponse', describeHostReservationsResponse_hostReservationSet - Details about the reservation's configuration.

$sel:httpStatus:DescribeHostReservationsResponse', describeHostReservationsResponse_httpStatus - The response's http status code.

Response Lenses

describeHostReservationsResponse_nextToken :: Lens' DescribeHostReservationsResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

describeHostReservationsResponse_hostReservationSet :: Lens' DescribeHostReservationsResponse (Maybe [HostReservation])

Details about the reservation's configuration.

describeHostReservationsResponse_httpStatus :: Lens' DescribeHostReservationsResponse Int

The response's http status code.