Network.AWS.EC2.DescribeIdentityIdFormat

Description

Describes the ID format settings for resources for the specified IAM user, IAM role, or root user. For example, you can view the resource types that are enabled for longer IDs. This request only returns information about resource types whose ID formats can be modified; it does not return information about other resource types. For more information, see Resource IDs in the Amazon Elastic Compute Cloud User Guide.

The following resource types support longer IDs: bundle | conversion-task | customer-gateway | dhcp-options | elastic-ip-allocation | elastic-ip-association | export-task | flow-log | image | import-task | instance | internet-gateway | network-acl | network-acl-association | network-interface | network-interface-attachment | prefix-list | reservation | route-table | route-table-association | security-group | snapshot | subnet | subnet-cidr-block-association | volume | vpc | vpc-cidr-block-association | vpc-endpoint | vpc-peering-connection | vpn-connection | vpn-gateway.

These settings apply to the principal specified in the request. They do not apply to the principal that makes the request.

Creating a Request

data DescribeIdentityIdFormat

See: newDescribeIdentityIdFormat smart constructor.

Constructors

DescribeIdentityIdFormat'

Fields

resource :: Maybe Text

The type of resource: bundle | conversion-task | customer-gateway | dhcp-options | elastic-ip-allocation | elastic-ip-association | export-task | flow-log | image | import-task | instance | internet-gateway | network-acl | network-acl-association | network-interface | network-interface-attachment | prefix-list | reservation | route-table | route-table-association | security-group | snapshot | subnet | subnet-cidr-block-association | volume | vpc | vpc-cidr-block-association | vpc-endpoint | vpc-peering-connection | vpn-connection | vpn-gateway

principalArn :: Text

The ARN of the principal, which can be an IAM role, IAM user, or the root user.

newDescribeIdentityIdFormat

Arguments

:: Text
$sel:principalArn:DescribeIdentityIdFormat'

-> DescribeIdentityIdFormat

Create a value of DescribeIdentityIdFormat with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resource:DescribeIdentityIdFormat', describeIdentityIdFormat_resource - The type of resource: bundle | conversion-task | customer-gateway | dhcp-options | elastic-ip-allocation | elastic-ip-association | export-task | flow-log | image | import-task | instance | internet-gateway | network-acl | network-acl-association | network-interface | network-interface-attachment | prefix-list | reservation | route-table | route-table-association | security-group | snapshot | subnet | subnet-cidr-block-association | volume | vpc | vpc-cidr-block-association | vpc-endpoint | vpc-peering-connection | vpn-connection | vpn-gateway

$sel:principalArn:DescribeIdentityIdFormat', describeIdentityIdFormat_principalArn - The ARN of the principal, which can be an IAM role, IAM user, or the root user.

Request Lenses

describeIdentityIdFormat_resource :: Lens' DescribeIdentityIdFormat (Maybe Text)

The type of resource: bundle | conversion-task | customer-gateway | dhcp-options | elastic-ip-allocation | elastic-ip-association | export-task | flow-log | image | import-task | instance | internet-gateway | network-acl | network-acl-association | network-interface | network-interface-attachment | prefix-list | reservation | route-table | route-table-association | security-group | snapshot | subnet | subnet-cidr-block-association | volume | vpc | vpc-cidr-block-association | vpc-endpoint | vpc-peering-connection | vpn-connection | vpn-gateway

describeIdentityIdFormat_principalArn :: Lens' DescribeIdentityIdFormat Text

The ARN of the principal, which can be an IAM role, IAM user, or the root user.

Destructuring the Response

data DescribeIdentityIdFormatResponse

See: newDescribeIdentityIdFormatResponse smart constructor.

Constructors

DescribeIdentityIdFormatResponse'

Fields

statuses :: Maybe [IdFormat]

Information about the ID format for the resources.

httpStatus :: Int

The response's http status code.

newDescribeIdentityIdFormatResponse

Arguments

:: Int
$sel:httpStatus:DescribeIdentityIdFormatResponse'

-> DescribeIdentityIdFormatResponse

Create a value of DescribeIdentityIdFormatResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:statuses:DescribeIdentityIdFormatResponse', describeIdentityIdFormatResponse_statuses - Information about the ID format for the resources.

$sel:httpStatus:DescribeIdentityIdFormatResponse', describeIdentityIdFormatResponse_httpStatus - The response's http status code.

Response Lenses

describeIdentityIdFormatResponse_statuses :: Lens' DescribeIdentityIdFormatResponse (Maybe [IdFormat])

Information about the ID format for the resources.

describeIdentityIdFormatResponse_httpStatus :: Lens' DescribeIdentityIdFormatResponse Int

The response's http status code.