Network.AWS.EC2.DescribeImageAttribute

Description

Describes the specified attribute of the specified AMI. You can specify only one attribute at a time.

Creating a Request

data DescribeImageAttribute

Contains the parameters for DescribeImageAttribute.

See: newDescribeImageAttribute smart constructor.

Constructors

DescribeImageAttribute'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

attribute :: ImageAttributeName

The AMI attribute.

Note: The blockDeviceMapping attribute is deprecated. Using this attribute returns the Client.AuthFailure error. To get information about the block device mappings for an AMI, use the DescribeImages action.

imageId :: Text

The ID of the AMI.

newDescribeImageAttribute

Arguments

:: ImageAttributeName
$sel:attribute:DescribeImageAttribute'

-> Text
$sel:imageId:DescribeImageAttribute'

-> DescribeImageAttribute

Create a value of DescribeImageAttribute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:DescribeImageAttribute', describeImageAttribute_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:attribute:DescribeImageAttribute', describeImageAttribute_attribute - The AMI attribute.

Note: The blockDeviceMapping attribute is deprecated. Using this attribute returns the Client.AuthFailure error. To get information about the block device mappings for an AMI, use the DescribeImages action.

$sel:imageId:DescribeImageAttribute', describeImageAttribute_imageId - The ID of the AMI.

Request Lenses

describeImageAttribute_dryRun :: Lens' DescribeImageAttribute (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeImageAttribute_attribute :: Lens' DescribeImageAttribute ImageAttributeName

The AMI attribute.

Note: The blockDeviceMapping attribute is deprecated. Using this attribute returns the Client.AuthFailure error. To get information about the block device mappings for an AMI, use the DescribeImages action.

describeImageAttribute_imageId :: Lens' DescribeImageAttribute Text

The ID of the AMI.

Destructuring the Response

data DescribeImageAttributeResponse

Describes an image attribute.

See: newDescribeImageAttributeResponse smart constructor.

Constructors

DescribeImageAttributeResponse'

Fields

ramdiskId :: Maybe AttributeValue

The RAM disk ID.

launchPermissions :: Maybe [LaunchPermission]

The launch permissions.

bootMode :: Maybe AttributeValue

productCodes :: Maybe [ProductCode]

The product codes.

imageId :: Maybe Text

The ID of the AMI.

sriovNetSupport :: Maybe AttributeValue

Indicates whether enhanced networking with the Intel 82599 Virtual Function interface is enabled.

blockDeviceMappings :: Maybe [BlockDeviceMapping]

The block device mapping entries.

description :: Maybe AttributeValue

A description for the AMI.

kernelId :: Maybe AttributeValue

The kernel ID.

httpStatus :: Int

The response's http status code.

newDescribeImageAttributeResponse

Arguments

:: Int
$sel:httpStatus:DescribeImageAttributeResponse'

-> DescribeImageAttributeResponse

Create a value of DescribeImageAttributeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ramdiskId:DescribeImageAttributeResponse', describeImageAttributeResponse_ramdiskId - The RAM disk ID.

$sel:launchPermissions:DescribeImageAttributeResponse', describeImageAttributeResponse_launchPermissions - The launch permissions.

$sel:bootMode:DescribeImageAttributeResponse', describeImageAttributeResponse_bootMode - Undocumented member.

$sel:productCodes:DescribeImageAttributeResponse', describeImageAttributeResponse_productCodes - The product codes.

$sel:imageId:DescribeImageAttribute', describeImageAttributeResponse_imageId - The ID of the AMI.

$sel:sriovNetSupport:DescribeImageAttributeResponse', describeImageAttributeResponse_sriovNetSupport - Indicates whether enhanced networking with the Intel 82599 Virtual Function interface is enabled.

$sel:blockDeviceMappings:DescribeImageAttributeResponse', describeImageAttributeResponse_blockDeviceMappings - The block device mapping entries.

$sel:description:DescribeImageAttributeResponse', describeImageAttributeResponse_description - A description for the AMI.

$sel:kernelId:DescribeImageAttributeResponse', describeImageAttributeResponse_kernelId - The kernel ID.

$sel:httpStatus:DescribeImageAttributeResponse', describeImageAttributeResponse_httpStatus - The response's http status code.

Response Lenses

describeImageAttributeResponse_ramdiskId :: Lens' DescribeImageAttributeResponse (Maybe AttributeValue)

The RAM disk ID.

describeImageAttributeResponse_launchPermissions :: Lens' DescribeImageAttributeResponse (Maybe [LaunchPermission])

The launch permissions.

describeImageAttributeResponse_bootMode :: Lens' DescribeImageAttributeResponse (Maybe AttributeValue)

Undocumented member.

describeImageAttributeResponse_productCodes :: Lens' DescribeImageAttributeResponse (Maybe [ProductCode])

The product codes.

describeImageAttributeResponse_imageId :: Lens' DescribeImageAttributeResponse (Maybe Text)

The ID of the AMI.

describeImageAttributeResponse_sriovNetSupport :: Lens' DescribeImageAttributeResponse (Maybe AttributeValue)

Indicates whether enhanced networking with the Intel 82599 Virtual Function interface is enabled.

describeImageAttributeResponse_blockDeviceMappings :: Lens' DescribeImageAttributeResponse (Maybe [BlockDeviceMapping])

The block device mapping entries.

describeImageAttributeResponse_description :: Lens' DescribeImageAttributeResponse (Maybe AttributeValue)

A description for the AMI.

describeImageAttributeResponse_kernelId :: Lens' DescribeImageAttributeResponse (Maybe AttributeValue)

The kernel ID.

describeImageAttributeResponse_httpStatus :: Lens' DescribeImageAttributeResponse Int

The response's http status code.