Network.AWS.EC2.DescribeImages

Description

Describes the specified images (AMIs, AKIs, and ARIs) available to you or all of the images available to you.

The images available to you include public images, private images that you own, and private images owned by other Amazon Web Services accounts for which you have explicit launch permissions.

Recently deregistered images appear in the returned results for a short interval and then return empty results. After all instances that reference a deregistered AMI are terminated, specifying the ID of the image will eventually return an error indicating that the AMI ID cannot be found.

Creating a Request

data DescribeImages

See: newDescribeImages smart constructor.

Constructors

DescribeImages'

Fields

imageIds :: Maybe [Text]

The image IDs.

Default: Describes all images available to you.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

owners :: Maybe [Text]

Scopes the results to images with the specified owners. You can specify a combination of Amazon Web Services account IDs, self, amazon, and aws-marketplace. If you omit this parameter, the results include all images for which you have launch permissions, regardless of ownership.

filters :: Maybe [Filter]

The filters.

architecture - The image architecture (i386 | x86_64 | arm64).

block-device-mapping.delete-on-termination - A Boolean value that indicates whether the Amazon EBS volume is deleted on instance termination.

block-device-mapping.device-name - The device name specified in the block device mapping (for example, /dev/sdh or xvdh).

block-device-mapping.snapshot-id - The ID of the snapshot used for the Amazon EBS volume.

block-device-mapping.volume-size - The volume size of the Amazon EBS volume, in GiB.

block-device-mapping.volume-type - The volume type of the Amazon EBS volume (io1 | io2 | gp2 | gp3 | sc1 | st1 | standard).

block-device-mapping.encrypted - A Boolean that indicates whether the Amazon EBS volume is encrypted.

description - The description of the image (provided during image creation).

ena-support - A Boolean that indicates whether enhanced networking with ENA is enabled.

hypervisor - The hypervisor type (ovm | xen).

image-id - The ID of the image.

image-type - The image type (machine | kernel | ramdisk).

is-public - A Boolean that indicates whether the image is public.

kernel-id - The kernel ID.

manifest-location - The location of the image manifest.

name - The name of the AMI (provided during image creation).

owner-alias - The owner alias (amazon | aws-marketplace). The valid aliases are defined in an Amazon-maintained list. This is not the Amazon Web Services account alias that can be set using the IAM console. We recommend that you use the Owner request parameter instead of this filter.

owner-id - The Amazon Web Services account ID of the owner. We recommend that you use the Owner request parameter instead of this filter.

platform - The platform. To only list Windows-based AMIs, use windows.

product-code - The product code.

product-code.type - The type of the product code (marketplace).

ramdisk-id - The RAM disk ID.

root-device-name - The device name of the root device volume (for example, /dev/sda1).

root-device-type - The type of the root device volume (ebs | instance-store).

state - The state of the image (available | pending | failed).

state-reason-code - The reason code for the state change.

state-reason-message - The message for the state change.

sriov-net-support - A value of simple indicates that enhanced networking with the Intel 82599 VF interface is enabled.

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

virtualization-type - The virtualization type (paravirtual | hvm).

includeDeprecated :: Maybe Bool

If true, all deprecated AMIs are included in the response. If false, no deprecated AMIs are included in the response. If no value is specified, the default value is false.

If you are the AMI owner, all deprecated AMIs appear in the response regardless of the value (true or false) that you set for this parameter.

executableUsers :: Maybe [Text]

Scopes the images by users with explicit launch permissions. Specify an Amazon Web Services account ID, self (the sender of the request), or all (public AMIs).

newDescribeImages :: DescribeImages

Create a value of DescribeImages with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:imageIds:DescribeImages', describeImages_imageIds - The image IDs.

Default: Describes all images available to you.

$sel:dryRun:DescribeImages', describeImages_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:owners:DescribeImages', describeImages_owners - Scopes the results to images with the specified owners. You can specify a combination of Amazon Web Services account IDs, self, amazon, and aws-marketplace. If you omit this parameter, the results include all images for which you have launch permissions, regardless of ownership.

$sel:filters:DescribeImages', describeImages_filters - The filters.

architecture - The image architecture (i386 | x86_64 | arm64).

block-device-mapping.delete-on-termination - A Boolean value that indicates whether the Amazon EBS volume is deleted on instance termination.

block-device-mapping.device-name - The device name specified in the block device mapping (for example, /dev/sdh or xvdh).

block-device-mapping.snapshot-id - The ID of the snapshot used for the Amazon EBS volume.

block-device-mapping.volume-size - The volume size of the Amazon EBS volume, in GiB.

block-device-mapping.volume-type - The volume type of the Amazon EBS volume (io1 | io2 | gp2 | gp3 | sc1 | st1 | standard).

block-device-mapping.encrypted - A Boolean that indicates whether the Amazon EBS volume is encrypted.

description - The description of the image (provided during image creation).

ena-support - A Boolean that indicates whether enhanced networking with ENA is enabled.

hypervisor - The hypervisor type (ovm | xen).

image-id - The ID of the image.

image-type - The image type (machine | kernel | ramdisk).

is-public - A Boolean that indicates whether the image is public.

kernel-id - The kernel ID.

manifest-location - The location of the image manifest.

name - The name of the AMI (provided during image creation).

owner-alias - The owner alias (amazon | aws-marketplace). The valid aliases are defined in an Amazon-maintained list. This is not the Amazon Web Services account alias that can be set using the IAM console. We recommend that you use the Owner request parameter instead of this filter.

owner-id - The Amazon Web Services account ID of the owner. We recommend that you use the Owner request parameter instead of this filter.

platform - The platform. To only list Windows-based AMIs, use windows.

product-code - The product code.

product-code.type - The type of the product code (marketplace).

ramdisk-id - The RAM disk ID.

root-device-name - The device name of the root device volume (for example, /dev/sda1).

root-device-type - The type of the root device volume (ebs | instance-store).

state - The state of the image (available | pending | failed).

state-reason-code - The reason code for the state change.

state-reason-message - The message for the state change.

sriov-net-support - A value of simple indicates that enhanced networking with the Intel 82599 VF interface is enabled.

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

virtualization-type - The virtualization type (paravirtual | hvm).

$sel:includeDeprecated:DescribeImages', describeImages_includeDeprecated - If true, all deprecated AMIs are included in the response. If false, no deprecated AMIs are included in the response. If no value is specified, the default value is false.

If you are the AMI owner, all deprecated AMIs appear in the response regardless of the value (true or false) that you set for this parameter.

$sel:executableUsers:DescribeImages', describeImages_executableUsers - Scopes the images by users with explicit launch permissions. Specify an Amazon Web Services account ID, self (the sender of the request), or all (public AMIs).

Request Lenses

describeImages_imageIds :: Lens' DescribeImages (Maybe [Text])

The image IDs.

Default: Describes all images available to you.

describeImages_dryRun :: Lens' DescribeImages (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeImages_owners :: Lens' DescribeImages (Maybe [Text])

Scopes the results to images with the specified owners. You can specify a combination of Amazon Web Services account IDs, self, amazon, and aws-marketplace. If you omit this parameter, the results include all images for which you have launch permissions, regardless of ownership.

describeImages_filters :: Lens' DescribeImages (Maybe [Filter])

The filters.

architecture - The image architecture (i386 | x86_64 | arm64).

block-device-mapping.delete-on-termination - A Boolean value that indicates whether the Amazon EBS volume is deleted on instance termination.

block-device-mapping.device-name - The device name specified in the block device mapping (for example, /dev/sdh or xvdh).

block-device-mapping.snapshot-id - The ID of the snapshot used for the Amazon EBS volume.

block-device-mapping.volume-size - The volume size of the Amazon EBS volume, in GiB.

block-device-mapping.volume-type - The volume type of the Amazon EBS volume (io1 | io2 | gp2 | gp3 | sc1 | st1 | standard).

block-device-mapping.encrypted - A Boolean that indicates whether the Amazon EBS volume is encrypted.

description - The description of the image (provided during image creation).

ena-support - A Boolean that indicates whether enhanced networking with ENA is enabled.

hypervisor - The hypervisor type (ovm | xen).

image-id - The ID of the image.

image-type - The image type (machine | kernel | ramdisk).

is-public - A Boolean that indicates whether the image is public.

kernel-id - The kernel ID.

manifest-location - The location of the image manifest.

name - The name of the AMI (provided during image creation).

owner-alias - The owner alias (amazon | aws-marketplace). The valid aliases are defined in an Amazon-maintained list. This is not the Amazon Web Services account alias that can be set using the IAM console. We recommend that you use the Owner request parameter instead of this filter.

owner-id - The Amazon Web Services account ID of the owner. We recommend that you use the Owner request parameter instead of this filter.

platform - The platform. To only list Windows-based AMIs, use windows.

product-code - The product code.

product-code.type - The type of the product code (marketplace).

ramdisk-id - The RAM disk ID.

root-device-name - The device name of the root device volume (for example, /dev/sda1).

root-device-type - The type of the root device volume (ebs | instance-store).

state - The state of the image (available | pending | failed).

state-reason-code - The reason code for the state change.

state-reason-message - The message for the state change.

sriov-net-support - A value of simple indicates that enhanced networking with the Intel 82599 VF interface is enabled.

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

virtualization-type - The virtualization type (paravirtual | hvm).

describeImages_includeDeprecated :: Lens' DescribeImages (Maybe Bool)

If true, all deprecated AMIs are included in the response. If false, no deprecated AMIs are included in the response. If no value is specified, the default value is false.

If you are the AMI owner, all deprecated AMIs appear in the response regardless of the value (true or false) that you set for this parameter.

describeImages_executableUsers :: Lens' DescribeImages (Maybe [Text])

Scopes the images by users with explicit launch permissions. Specify an Amazon Web Services account ID, self (the sender of the request), or all (public AMIs).

Destructuring the Response

data DescribeImagesResponse

See: newDescribeImagesResponse smart constructor.

Constructors

DescribeImagesResponse'

Fields

images :: Maybe [Image]

Information about the images.

httpStatus :: Int

The response's http status code.

newDescribeImagesResponse

Arguments

:: Int
$sel:httpStatus:DescribeImagesResponse'

-> DescribeImagesResponse

Create a value of DescribeImagesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:images:DescribeImagesResponse', describeImagesResponse_images - Information about the images.

$sel:httpStatus:DescribeImagesResponse', describeImagesResponse_httpStatus - The response's http status code.

Response Lenses

describeImagesResponse_images :: Lens' DescribeImagesResponse (Maybe [Image])

Information about the images.

describeImagesResponse_httpStatus :: Lens' DescribeImagesResponse Int

The response's http status code.