Network.AWS.EC2.DescribeImportImageTasks

Description

Displays details about an import virtual machine or import snapshot tasks that are already created.

This operation returns paginated results.

Creating a Request

data DescribeImportImageTasks

See: newDescribeImportImageTasks smart constructor.

Constructors

DescribeImportImageTasks'

Fields

nextToken :: Maybe Text

A token that indicates the next page of results.

importTaskIds :: Maybe [Text]

The IDs of the import image tasks.

maxResults :: Maybe Int

The maximum number of results to return in a single call.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

filters :: Maybe [Filter]

Filter tasks using the task-state filter and one of the following values: active, completed, deleting, or deleted.

newDescribeImportImageTasks :: DescribeImportImageTasks

Create a value of DescribeImportImageTasks with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeImportImageTasks', describeImportImageTasks_nextToken - A token that indicates the next page of results.

$sel:importTaskIds:DescribeImportImageTasks', describeImportImageTasks_importTaskIds - The IDs of the import image tasks.

$sel:maxResults:DescribeImportImageTasks', describeImportImageTasks_maxResults - The maximum number of results to return in a single call.

$sel:dryRun:DescribeImportImageTasks', describeImportImageTasks_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:filters:DescribeImportImageTasks', describeImportImageTasks_filters - Filter tasks using the task-state filter and one of the following values: active, completed, deleting, or deleted.

Request Lenses

describeImportImageTasks_nextToken :: Lens' DescribeImportImageTasks (Maybe Text)

A token that indicates the next page of results.

describeImportImageTasks_importTaskIds :: Lens' DescribeImportImageTasks (Maybe [Text])

The IDs of the import image tasks.

describeImportImageTasks_maxResults :: Lens' DescribeImportImageTasks (Maybe Int)

The maximum number of results to return in a single call.

describeImportImageTasks_dryRun :: Lens' DescribeImportImageTasks (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeImportImageTasks_filters :: Lens' DescribeImportImageTasks (Maybe [Filter])

Filter tasks using the task-state filter and one of the following values: active, completed, deleting, or deleted.

Destructuring the Response

data DescribeImportImageTasksResponse

See: newDescribeImportImageTasksResponse smart constructor.

Constructors

DescribeImportImageTasksResponse'

Fields

nextToken :: Maybe Text

The token to use to get the next page of results. This value is null when there are no more results to return.

importImageTasks :: Maybe [ImportImageTask]

A list of zero or more import image tasks that are currently active or were completed or canceled in the previous 7 days.

httpStatus :: Int

The response's http status code.

newDescribeImportImageTasksResponse

Arguments

:: Int
$sel:httpStatus:DescribeImportImageTasksResponse'

-> DescribeImportImageTasksResponse

Create a value of DescribeImportImageTasksResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeImportImageTasks', describeImportImageTasksResponse_nextToken - The token to use to get the next page of results. This value is null when there are no more results to return.

$sel:importImageTasks:DescribeImportImageTasksResponse', describeImportImageTasksResponse_importImageTasks - A list of zero or more import image tasks that are currently active or were completed or canceled in the previous 7 days.

$sel:httpStatus:DescribeImportImageTasksResponse', describeImportImageTasksResponse_httpStatus - The response's http status code.

Response Lenses

describeImportImageTasksResponse_nextToken :: Lens' DescribeImportImageTasksResponse (Maybe Text)

The token to use to get the next page of results. This value is null when there are no more results to return.

describeImportImageTasksResponse_importImageTasks :: Lens' DescribeImportImageTasksResponse (Maybe [ImportImageTask])

A list of zero or more import image tasks that are currently active or were completed or canceled in the previous 7 days.

describeImportImageTasksResponse_httpStatus :: Lens' DescribeImportImageTasksResponse Int

The response's http status code.