Network.AWS.EC2.DescribeImportSnapshotTasks

Description

Describes your import snapshot tasks.

This operation returns paginated results.

Creating a Request

data DescribeImportSnapshotTasks

See: newDescribeImportSnapshotTasks smart constructor.

Constructors

DescribeImportSnapshotTasks'

Fields

filters :: Maybe [Filter]

The filters.

importTaskIds :: Maybe [Text]

A list of import snapshot task IDs.

nextToken :: Maybe Text

A token that indicates the next page of results.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

maxResults :: Maybe Int

The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned NextToken value.

newDescribeImportSnapshotTasks :: DescribeImportSnapshotTasks

Create a value of DescribeImportSnapshotTasks with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:DescribeImportSnapshotTasks', describeImportSnapshotTasks_filters - The filters.

$sel:importTaskIds:DescribeImportSnapshotTasks', describeImportSnapshotTasks_importTaskIds - A list of import snapshot task IDs.

$sel:nextToken:DescribeImportSnapshotTasks', describeImportSnapshotTasks_nextToken - A token that indicates the next page of results.

$sel:dryRun:DescribeImportSnapshotTasks', describeImportSnapshotTasks_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:maxResults:DescribeImportSnapshotTasks', describeImportSnapshotTasks_maxResults - The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned NextToken value.

Request Lenses

describeImportSnapshotTasks_filters :: Lens' DescribeImportSnapshotTasks (Maybe [Filter])

The filters.

describeImportSnapshotTasks_importTaskIds :: Lens' DescribeImportSnapshotTasks (Maybe [Text])

A list of import snapshot task IDs.

describeImportSnapshotTasks_nextToken :: Lens' DescribeImportSnapshotTasks (Maybe Text)

A token that indicates the next page of results.

describeImportSnapshotTasks_dryRun :: Lens' DescribeImportSnapshotTasks (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeImportSnapshotTasks_maxResults :: Lens' DescribeImportSnapshotTasks (Maybe Int)

The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned NextToken value.

Destructuring the Response

data DescribeImportSnapshotTasksResponse

See: newDescribeImportSnapshotTasksResponse smart constructor.

Constructors

DescribeImportSnapshotTasksResponse'

Fields

nextToken :: Maybe Text

The token to use to get the next page of results. This value is null when there are no more results to return.

importSnapshotTasks :: Maybe [ImportSnapshotTask]

A list of zero or more import snapshot tasks that are currently active or were completed or canceled in the previous 7 days.

httpStatus :: Int

The response's http status code.

newDescribeImportSnapshotTasksResponse

Arguments

:: Int
$sel:httpStatus:DescribeImportSnapshotTasksResponse'

-> DescribeImportSnapshotTasksResponse

Create a value of DescribeImportSnapshotTasksResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeImportSnapshotTasks', describeImportSnapshotTasksResponse_nextToken - The token to use to get the next page of results. This value is null when there are no more results to return.

$sel:importSnapshotTasks:DescribeImportSnapshotTasksResponse', describeImportSnapshotTasksResponse_importSnapshotTasks - A list of zero or more import snapshot tasks that are currently active or were completed or canceled in the previous 7 days.

$sel:httpStatus:DescribeImportSnapshotTasksResponse', describeImportSnapshotTasksResponse_httpStatus - The response's http status code.

Response Lenses

describeImportSnapshotTasksResponse_nextToken :: Lens' DescribeImportSnapshotTasksResponse (Maybe Text)

The token to use to get the next page of results. This value is null when there are no more results to return.

describeImportSnapshotTasksResponse_importSnapshotTasks :: Lens' DescribeImportSnapshotTasksResponse (Maybe [ImportSnapshotTask])

A list of zero or more import snapshot tasks that are currently active or were completed or canceled in the previous 7 days.

describeImportSnapshotTasksResponse_httpStatus :: Lens' DescribeImportSnapshotTasksResponse Int

The response's http status code.