Network.AWS.EC2.DescribeInstanceEventWindows

Description

Describes the specified event windows or all event windows.

If you specify event window IDs, the output includes information for only the specified event windows. If you specify filters, the output includes information for only those event windows that meet the filter criteria. If you do not specify event windows IDs or filters, the output includes information for all event windows, which can affect performance. We recommend that you use pagination to ensure that the operation returns quickly and successfully.

For more information, see Define event windows for scheduled events in the Amazon EC2 User Guide.

This operation returns paginated results.

Creating a Request

data DescribeInstanceEventWindows

Describe instance event windows by InstanceEventWindow.

See: newDescribeInstanceEventWindows smart constructor.

Constructors

DescribeInstanceEventWindows'

Fields

instanceEventWindowIds :: Maybe [Text]

The IDs of the event windows.

nextToken :: Maybe Text

The token to request the next page of results.

maxResults :: Maybe Natural

The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned NextToken value. This value can be between 20 and 500. You cannot specify this parameter and the event window IDs parameter in the same call.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

filters :: Maybe [Filter]

One or more filters.

dedicated-host-id - The event windows associated with the specified Dedicated Host ID.

event-window-name - The event windows associated with the specified names.

instance-id - The event windows associated with the specified instance ID.

instance-tag - The event windows associated with the specified tag and value.

instance-tag-key - The event windows associated with the specified tag key, regardless of the value.

instance-tag-value - The event windows associated with the specified tag value, regardless of the key.

tag:<key> - The key/value combination of a tag assigned to the event window. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value CMX, specify tag:Owner for the filter name and CMX for the filter value.

tag-key - The key of a tag assigned to the event window. Use this filter to find all event windows that have a tag with a specific key, regardless of the tag value.

tag-value - The value of a tag assigned to the event window. Use this filter to find all event windows that have a tag with a specific value, regardless of the tag key.

newDescribeInstanceEventWindows :: DescribeInstanceEventWindows

Create a value of DescribeInstanceEventWindows with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceEventWindowIds:DescribeInstanceEventWindows', describeInstanceEventWindows_instanceEventWindowIds - The IDs of the event windows.

$sel:nextToken:DescribeInstanceEventWindows', describeInstanceEventWindows_nextToken - The token to request the next page of results.

$sel:maxResults:DescribeInstanceEventWindows', describeInstanceEventWindows_maxResults - The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned NextToken value. This value can be between 20 and 500. You cannot specify this parameter and the event window IDs parameter in the same call.

$sel:dryRun:DescribeInstanceEventWindows', describeInstanceEventWindows_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:filters:DescribeInstanceEventWindows', describeInstanceEventWindows_filters - One or more filters.

dedicated-host-id - The event windows associated with the specified Dedicated Host ID.

event-window-name - The event windows associated with the specified names.

instance-id - The event windows associated with the specified instance ID.

instance-tag - The event windows associated with the specified tag and value.

instance-tag-key - The event windows associated with the specified tag key, regardless of the value.

instance-tag-value - The event windows associated with the specified tag value, regardless of the key.

tag:<key> - The key/value combination of a tag assigned to the event window. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value CMX, specify tag:Owner for the filter name and CMX for the filter value.

tag-key - The key of a tag assigned to the event window. Use this filter to find all event windows that have a tag with a specific key, regardless of the tag value.

tag-value - The value of a tag assigned to the event window. Use this filter to find all event windows that have a tag with a specific value, regardless of the tag key.

Request Lenses

describeInstanceEventWindows_instanceEventWindowIds :: Lens' DescribeInstanceEventWindows (Maybe [Text])

The IDs of the event windows.

describeInstanceEventWindows_nextToken :: Lens' DescribeInstanceEventWindows (Maybe Text)

The token to request the next page of results.

describeInstanceEventWindows_maxResults :: Lens' DescribeInstanceEventWindows (Maybe Natural)

The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned NextToken value. This value can be between 20 and 500. You cannot specify this parameter and the event window IDs parameter in the same call.

describeInstanceEventWindows_dryRun :: Lens' DescribeInstanceEventWindows (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeInstanceEventWindows_filters :: Lens' DescribeInstanceEventWindows (Maybe [Filter])

One or more filters.

dedicated-host-id - The event windows associated with the specified Dedicated Host ID.

event-window-name - The event windows associated with the specified names.

instance-id - The event windows associated with the specified instance ID.

instance-tag - The event windows associated with the specified tag and value.

instance-tag-key - The event windows associated with the specified tag key, regardless of the value.

instance-tag-value - The event windows associated with the specified tag value, regardless of the key.

tag:<key> - The key/value combination of a tag assigned to the event window. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value CMX, specify tag:Owner for the filter name and CMX for the filter value.

tag-key - The key of a tag assigned to the event window. Use this filter to find all event windows that have a tag with a specific key, regardless of the tag value.

tag-value - The value of a tag assigned to the event window. Use this filter to find all event windows that have a tag with a specific value, regardless of the tag key.

Destructuring the Response

data DescribeInstanceEventWindowsResponse

See: newDescribeInstanceEventWindowsResponse smart constructor.

Constructors

DescribeInstanceEventWindowsResponse'

Fields

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

instanceEventWindows :: Maybe [InstanceEventWindow]

Information about the event windows.

httpStatus :: Int

The response's http status code.

newDescribeInstanceEventWindowsResponse

Arguments

:: Int
$sel:httpStatus:DescribeInstanceEventWindowsResponse'

-> DescribeInstanceEventWindowsResponse

Create a value of DescribeInstanceEventWindowsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeInstanceEventWindows', describeInstanceEventWindowsResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:instanceEventWindows:DescribeInstanceEventWindowsResponse', describeInstanceEventWindowsResponse_instanceEventWindows - Information about the event windows.

$sel:httpStatus:DescribeInstanceEventWindowsResponse', describeInstanceEventWindowsResponse_httpStatus - The response's http status code.

Response Lenses

describeInstanceEventWindowsResponse_nextToken :: Lens' DescribeInstanceEventWindowsResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

describeInstanceEventWindowsResponse_instanceEventWindows :: Lens' DescribeInstanceEventWindowsResponse (Maybe [InstanceEventWindow])

Information about the event windows.

describeInstanceEventWindowsResponse_httpStatus :: Lens' DescribeInstanceEventWindowsResponse Int

The response's http status code.