Network.AWS.EC2.DescribeInstanceStatus

Description

Describes the status of the specified instances or all of your instances. By default, only running instances are described, unless you specifically indicate to return the status of all instances.

Instance status includes the following components:

Status checks - Amazon EC2 performs status checks on running EC2 instances to identify hardware and software issues. For more information, see Status checks for your instances and Troubleshooting instances with failed status checks in the Amazon EC2 User Guide.

Scheduled events - Amazon EC2 can schedule events (such as reboot, stop, or terminate) for your instances related to hardware issues, software updates, or system maintenance. For more information, see Scheduled events for your instances in the Amazon EC2 User Guide.

Instance state - You can manage your instances from the moment you launch them through their termination. For more information, see Instance lifecycle in the Amazon EC2 User Guide.

This operation returns paginated results.

Creating a Request

data DescribeInstanceStatus

See: newDescribeInstanceStatus smart constructor.

Constructors

DescribeInstanceStatus'

Fields

includeAllInstances :: Maybe Bool

When true, includes the health status for all instances. When false, includes the health status for running instances only.

Default: false

filters :: Maybe [Filter]

The filters.

availability-zone - The Availability Zone of the instance.

event.code - The code for the scheduled event (instance-reboot | system-reboot | system-maintenance | instance-retirement | instance-stop).

event.description - A description of the event.

event.instance-event-id - The ID of the event whose date and time you are modifying.

event.not-after - The latest end time for the scheduled event (for example, 2014-09-15T17:15:20.000Z).

event.not-before - The earliest start time for the scheduled event (for example, 2014-09-15T17:15:20.000Z).

event.not-before-deadline - The deadline for starting the event (for example, 2014-09-15T17:15:20.000Z).

instance-state-code - The code for the instance state, as a 16-bit unsigned integer. The high byte is used for internal purposes and should be ignored. The low byte is set based on the state represented. The valid values are 0 (pending), 16 (running), 32 (shutting-down), 48 (terminated), 64 (stopping), and 80 (stopped).

instance-state-name - The state of the instance (pending | running | shutting-down | terminated | stopping | stopped).

instance-status.reachability - Filters on instance status where the name is reachability (passed | failed | initializing | insufficient-data).

instance-status.status - The status of the instance (ok | impaired | initializing | insufficient-data | not-applicable).

system-status.reachability - Filters on system status where the name is reachability (passed | failed | initializing | insufficient-data).

system-status.status - The system status of the instance (ok | impaired | initializing | insufficient-data | not-applicable).

nextToken :: Maybe Text

The token to retrieve the next page of results.

instanceIds :: Maybe [Text]

The instance IDs.

Default: Describes all your instances.

Constraints: Maximum 100 explicitly specified instance IDs.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

maxResults :: Maybe Int

The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned NextToken value. This value can be between 5 and 1000. You cannot specify this parameter and the instance IDs parameter in the same call.

newDescribeInstanceStatus :: DescribeInstanceStatus

Create a value of DescribeInstanceStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:includeAllInstances:DescribeInstanceStatus', describeInstanceStatus_includeAllInstances - When true, includes the health status for all instances. When false, includes the health status for running instances only.

Default: false

$sel:filters:DescribeInstanceStatus', describeInstanceStatus_filters - The filters.

availability-zone - The Availability Zone of the instance.

event.code - The code for the scheduled event (instance-reboot | system-reboot | system-maintenance | instance-retirement | instance-stop).

event.description - A description of the event.

event.instance-event-id - The ID of the event whose date and time you are modifying.

event.not-after - The latest end time for the scheduled event (for example, 2014-09-15T17:15:20.000Z).

event.not-before - The earliest start time for the scheduled event (for example, 2014-09-15T17:15:20.000Z).

event.not-before-deadline - The deadline for starting the event (for example, 2014-09-15T17:15:20.000Z).

instance-state-code - The code for the instance state, as a 16-bit unsigned integer. The high byte is used for internal purposes and should be ignored. The low byte is set based on the state represented. The valid values are 0 (pending), 16 (running), 32 (shutting-down), 48 (terminated), 64 (stopping), and 80 (stopped).

instance-state-name - The state of the instance (pending | running | shutting-down | terminated | stopping | stopped).

instance-status.reachability - Filters on instance status where the name is reachability (passed | failed | initializing | insufficient-data).

instance-status.status - The status of the instance (ok | impaired | initializing | insufficient-data | not-applicable).

system-status.reachability - Filters on system status where the name is reachability (passed | failed | initializing | insufficient-data).

system-status.status - The system status of the instance (ok | impaired | initializing | insufficient-data | not-applicable).

$sel:nextToken:DescribeInstanceStatus', describeInstanceStatus_nextToken - The token to retrieve the next page of results.

$sel:instanceIds:DescribeInstanceStatus', describeInstanceStatus_instanceIds - The instance IDs.

Default: Describes all your instances.

Constraints: Maximum 100 explicitly specified instance IDs.

$sel:dryRun:DescribeInstanceStatus', describeInstanceStatus_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:maxResults:DescribeInstanceStatus', describeInstanceStatus_maxResults - The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned NextToken value. This value can be between 5 and 1000. You cannot specify this parameter and the instance IDs parameter in the same call.

Request Lenses

describeInstanceStatus_includeAllInstances :: Lens' DescribeInstanceStatus (Maybe Bool)

When true, includes the health status for all instances. When false, includes the health status for running instances only.

Default: false

describeInstanceStatus_filters :: Lens' DescribeInstanceStatus (Maybe [Filter])

The filters.

availability-zone - The Availability Zone of the instance.

event.code - The code for the scheduled event (instance-reboot | system-reboot | system-maintenance | instance-retirement | instance-stop).

event.description - A description of the event.

event.instance-event-id - The ID of the event whose date and time you are modifying.

event.not-after - The latest end time for the scheduled event (for example, 2014-09-15T17:15:20.000Z).

event.not-before - The earliest start time for the scheduled event (for example, 2014-09-15T17:15:20.000Z).

event.not-before-deadline - The deadline for starting the event (for example, 2014-09-15T17:15:20.000Z).

instance-state-code - The code for the instance state, as a 16-bit unsigned integer. The high byte is used for internal purposes and should be ignored. The low byte is set based on the state represented. The valid values are 0 (pending), 16 (running), 32 (shutting-down), 48 (terminated), 64 (stopping), and 80 (stopped).

instance-state-name - The state of the instance (pending | running | shutting-down | terminated | stopping | stopped).

instance-status.reachability - Filters on instance status where the name is reachability (passed | failed | initializing | insufficient-data).

instance-status.status - The status of the instance (ok | impaired | initializing | insufficient-data | not-applicable).

system-status.reachability - Filters on system status where the name is reachability (passed | failed | initializing | insufficient-data).

system-status.status - The system status of the instance (ok | impaired | initializing | insufficient-data | not-applicable).

describeInstanceStatus_nextToken :: Lens' DescribeInstanceStatus (Maybe Text)

The token to retrieve the next page of results.

describeInstanceStatus_instanceIds :: Lens' DescribeInstanceStatus (Maybe [Text])

The instance IDs.

Default: Describes all your instances.

Constraints: Maximum 100 explicitly specified instance IDs.

describeInstanceStatus_dryRun :: Lens' DescribeInstanceStatus (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeInstanceStatus_maxResults :: Lens' DescribeInstanceStatus (Maybe Int)

The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned NextToken value. This value can be between 5 and 1000. You cannot specify this parameter and the instance IDs parameter in the same call.

Destructuring the Response

data DescribeInstanceStatusResponse

See: newDescribeInstanceStatusResponse smart constructor.

Constructors

DescribeInstanceStatusResponse'

Fields

instanceStatuses :: Maybe [InstanceStatus]

Information about the status of the instances.

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

httpStatus :: Int

The response's http status code.

newDescribeInstanceStatusResponse

Arguments

:: Int
$sel:httpStatus:DescribeInstanceStatusResponse'

-> DescribeInstanceStatusResponse

Create a value of DescribeInstanceStatusResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceStatuses:DescribeInstanceStatusResponse', describeInstanceStatusResponse_instanceStatuses - Information about the status of the instances.

$sel:nextToken:DescribeInstanceStatus', describeInstanceStatusResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:httpStatus:DescribeInstanceStatusResponse', describeInstanceStatusResponse_httpStatus - The response's http status code.

Response Lenses

describeInstanceStatusResponse_instanceStatuses :: Lens' DescribeInstanceStatusResponse (Maybe [InstanceStatus])

Information about the status of the instances.

describeInstanceStatusResponse_nextToken :: Lens' DescribeInstanceStatusResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

describeInstanceStatusResponse_httpStatus :: Lens' DescribeInstanceStatusResponse Int

The response's http status code.