Network.AWS.EC2.DescribeInstanceTypeOfferings

Description

Returns a list of all instance types offered. The results can be filtered by location (Region or Availability Zone). If no location is specified, the instance types offered in the current Region are returned.

This operation returns paginated results.

Creating a Request

data DescribeInstanceTypeOfferings

See: newDescribeInstanceTypeOfferings smart constructor.

Constructors

DescribeInstanceTypeOfferings'

Fields

filters :: Maybe [Filter]

One or more filters. Filter names and values are case-sensitive.

location - This depends on the location type. For example, if the location type is region (default), the location is the Region code (for example, us-east-2.)

instance-type - The instance type. For example, c5.2xlarge.

nextToken :: Maybe Text

The token to retrieve the next page of results.

locationType :: Maybe LocationType

The location type.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

maxResults :: Maybe Natural

The maximum number of results to return for the request in a single page. The remaining results can be seen by sending another request with the next token value.

newDescribeInstanceTypeOfferings :: DescribeInstanceTypeOfferings

Create a value of DescribeInstanceTypeOfferings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:DescribeInstanceTypeOfferings', describeInstanceTypeOfferings_filters - One or more filters. Filter names and values are case-sensitive.

location - This depends on the location type. For example, if the location type is region (default), the location is the Region code (for example, us-east-2.)

instance-type - The instance type. For example, c5.2xlarge.

$sel:nextToken:DescribeInstanceTypeOfferings', describeInstanceTypeOfferings_nextToken - The token to retrieve the next page of results.

$sel:locationType:DescribeInstanceTypeOfferings', describeInstanceTypeOfferings_locationType - The location type.

$sel:dryRun:DescribeInstanceTypeOfferings', describeInstanceTypeOfferings_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:maxResults:DescribeInstanceTypeOfferings', describeInstanceTypeOfferings_maxResults - The maximum number of results to return for the request in a single page. The remaining results can be seen by sending another request with the next token value.

Request Lenses

describeInstanceTypeOfferings_filters :: Lens' DescribeInstanceTypeOfferings (Maybe [Filter])

One or more filters. Filter names and values are case-sensitive.

location - This depends on the location type. For example, if the location type is region (default), the location is the Region code (for example, us-east-2.)

instance-type - The instance type. For example, c5.2xlarge.

describeInstanceTypeOfferings_nextToken :: Lens' DescribeInstanceTypeOfferings (Maybe Text)

The token to retrieve the next page of results.

describeInstanceTypeOfferings_locationType :: Lens' DescribeInstanceTypeOfferings (Maybe LocationType)

The location type.

describeInstanceTypeOfferings_dryRun :: Lens' DescribeInstanceTypeOfferings (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeInstanceTypeOfferings_maxResults :: Lens' DescribeInstanceTypeOfferings (Maybe Natural)

The maximum number of results to return for the request in a single page. The remaining results can be seen by sending another request with the next token value.

Destructuring the Response

data DescribeInstanceTypeOfferingsResponse

See: newDescribeInstanceTypeOfferingsResponse smart constructor.

Constructors

DescribeInstanceTypeOfferingsResponse'

Fields

instanceTypeOfferings :: Maybe [InstanceTypeOffering]

The instance types offered.

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

httpStatus :: Int

The response's http status code.

newDescribeInstanceTypeOfferingsResponse

Arguments

:: Int
$sel:httpStatus:DescribeInstanceTypeOfferingsResponse'

-> DescribeInstanceTypeOfferingsResponse

Create a value of DescribeInstanceTypeOfferingsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceTypeOfferings:DescribeInstanceTypeOfferingsResponse', describeInstanceTypeOfferingsResponse_instanceTypeOfferings - The instance types offered.

$sel:nextToken:DescribeInstanceTypeOfferings', describeInstanceTypeOfferingsResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:httpStatus:DescribeInstanceTypeOfferingsResponse', describeInstanceTypeOfferingsResponse_httpStatus - The response's http status code.

Response Lenses

describeInstanceTypeOfferingsResponse_instanceTypeOfferings :: Lens' DescribeInstanceTypeOfferingsResponse (Maybe [InstanceTypeOffering])

The instance types offered.

describeInstanceTypeOfferingsResponse_nextToken :: Lens' DescribeInstanceTypeOfferingsResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

describeInstanceTypeOfferingsResponse_httpStatus :: Lens' DescribeInstanceTypeOfferingsResponse Int

The response's http status code.