Network.AWS.EC2.DescribeInstanceTypes

Description

Describes the details of the instance types that are offered in a location. The results can be filtered by the attributes of the instance types.

This operation returns paginated results.

Creating a Request

data DescribeInstanceTypes

See: newDescribeInstanceTypes smart constructor.

Constructors

DescribeInstanceTypes'

Fields

nextToken :: Maybe Text

The token to retrieve the next page of results.

maxResults :: Maybe Natural

The maximum number of results to return for the request in a single page. The remaining results can be seen by sending another request with the next token value.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

filters :: Maybe [Filter]

One or more filters. Filter names and values are case-sensitive.

auto-recovery-supported - Indicates whether auto recovery is supported (true | false).

bare-metal - Indicates whether it is a bare metal instance type (true | false).

burstable-performance-supported - Indicates whether it is a burstable performance instance type (true | false).

current-generation - Indicates whether this instance type is the latest generation instance type of an instance family (true | false).

ebs-info.ebs-optimized-info.baseline-bandwidth-in-mbps - The baseline bandwidth performance for an EBS-optimized instance type, in Mbps.

ebs-info.ebs-optimized-info.baseline-iops - The baseline input/output storage operations per second for an EBS-optimized instance type.

ebs-info.ebs-optimized-info.baseline-throughput-in-mbps - The baseline throughput performance for an EBS-optimized instance type, in MB/s.

ebs-info.ebs-optimized-info.maximum-bandwidth-in-mbps - The maximum bandwidth performance for an EBS-optimized instance type, in Mbps.

ebs-info.ebs-optimized-info.maximum-iops - The maximum input/output storage operations per second for an EBS-optimized instance type.

ebs-info.ebs-optimized-info.maximum-throughput-in-mbps - The maximum throughput performance for an EBS-optimized instance type, in MB/s.

ebs-info.ebs-optimized-support - Indicates whether the instance type is EBS-optimized (supported | unsupported | default).

ebs-info.encryption-support - Indicates whether EBS encryption is supported (supported | unsupported).

ebs-info.nvme-support - Indicates whether non-volatile memory express (NVMe) is supported for EBS volumes (required | supported | unsupported).

free-tier-eligible - Indicates whether the instance type is eligible to use in the free tier (true | false).

hibernation-supported - Indicates whether On-Demand hibernation is supported (true | false).

hypervisor - The hypervisor (nitro | xen).

instance-storage-info.disk.count - The number of local disks.

instance-storage-info.disk.size-in-gb - The storage size of each instance storage disk, in GB.

instance-storage-info.disk.type - The storage technology for the local instance storage disks (hdd | ssd).

instance-storage-info.nvme-support - Indicates whether non-volatile memory express (NVMe) is supported for instance store (required | supported) | unsupported).

instance-storage-info.total-size-in-gb - The total amount of storage available from all local instance storage, in GB.

instance-storage-supported - Indicates whether the instance type has local instance storage (true | false).

instance-type - The instance type (for example c5.2xlarge or c5*).

memory-info.size-in-mib - The memory size.

network-info.efa-info.maximum-efa-interfaces - The maximum number of Elastic Fabric Adapters (EFAs) per instance.

network-info.efa-supported - Indicates whether the instance type supports Elastic Fabric Adapter (EFA) (true | false).

network-info.ena-support - Indicates whether Elastic Network Adapter (ENA) is supported or required (required | supported | unsupported).

network-info.encryption-in-transit-supported - Indicates whether the instance type automatically encrypts in-transit traffic between instances (true | false).

network-info.ipv4-addresses-per-interface - The maximum number of private IPv4 addresses per network interface.

network-info.ipv6-addresses-per-interface - The maximum number of private IPv6 addresses per network interface.

network-info.ipv6-supported - Indicates whether the instance type supports IPv6 (true | false).

network-info.maximum-network-interfaces - The maximum number of network interfaces per instance.

network-info.network-performance - The network performance (for example, "25 Gigabit").

processor-info.supported-architecture - The CPU architecture (arm64 | i386 | x86_64).

processor-info.sustained-clock-speed-in-ghz - The CPU clock speed, in GHz.

supported-boot-mode - The boot mode (legacy-bios | uefi).

supported-root-device-type - The root device type (ebs | instance-store).

supported-usage-class - The usage class (on-demand | spot).

supported-virtualization-type - The virtualization type (hvm | paravirtual).

vcpu-info.default-cores - The default number of cores for the instance type.

vcpu-info.default-threads-per-core - The default number of threads per core for the instance type.

vcpu-info.default-vcpus - The default number of vCPUs for the instance type.

vcpu-info.valid-cores - The number of cores that can be configured for the instance type.

vcpu-info.valid-threads-per-core - The number of threads per core that can be configured for the instance type. For example, "1" or "1,2".

instanceTypes :: Maybe [InstanceType]

The instance types. For more information, see Instance types in the Amazon EC2 User Guide.

newDescribeInstanceTypes :: DescribeInstanceTypes

Create a value of DescribeInstanceTypes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeInstanceTypes', describeInstanceTypes_nextToken - The token to retrieve the next page of results.

$sel:maxResults:DescribeInstanceTypes', describeInstanceTypes_maxResults - The maximum number of results to return for the request in a single page. The remaining results can be seen by sending another request with the next token value.

$sel:dryRun:DescribeInstanceTypes', describeInstanceTypes_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:filters:DescribeInstanceTypes', describeInstanceTypes_filters - One or more filters. Filter names and values are case-sensitive.

auto-recovery-supported - Indicates whether auto recovery is supported (true | false).

bare-metal - Indicates whether it is a bare metal instance type (true | false).

burstable-performance-supported - Indicates whether it is a burstable performance instance type (true | false).

current-generation - Indicates whether this instance type is the latest generation instance type of an instance family (true | false).

ebs-info.ebs-optimized-info.baseline-bandwidth-in-mbps - The baseline bandwidth performance for an EBS-optimized instance type, in Mbps.

ebs-info.ebs-optimized-info.baseline-iops - The baseline input/output storage operations per second for an EBS-optimized instance type.

ebs-info.ebs-optimized-info.baseline-throughput-in-mbps - The baseline throughput performance for an EBS-optimized instance type, in MB/s.

ebs-info.ebs-optimized-info.maximum-bandwidth-in-mbps - The maximum bandwidth performance for an EBS-optimized instance type, in Mbps.

ebs-info.ebs-optimized-info.maximum-iops - The maximum input/output storage operations per second for an EBS-optimized instance type.

ebs-info.ebs-optimized-info.maximum-throughput-in-mbps - The maximum throughput performance for an EBS-optimized instance type, in MB/s.

ebs-info.ebs-optimized-support - Indicates whether the instance type is EBS-optimized (supported | unsupported | default).

ebs-info.encryption-support - Indicates whether EBS encryption is supported (supported | unsupported).

ebs-info.nvme-support - Indicates whether non-volatile memory express (NVMe) is supported for EBS volumes (required | supported | unsupported).

free-tier-eligible - Indicates whether the instance type is eligible to use in the free tier (true | false).

hibernation-supported - Indicates whether On-Demand hibernation is supported (true | false).

hypervisor - The hypervisor (nitro | xen).

instance-storage-info.disk.count - The number of local disks.

instance-storage-info.disk.size-in-gb - The storage size of each instance storage disk, in GB.

instance-storage-info.disk.type - The storage technology for the local instance storage disks (hdd | ssd).

instance-storage-info.nvme-support - Indicates whether non-volatile memory express (NVMe) is supported for instance store (required | supported) | unsupported).

instance-storage-info.total-size-in-gb - The total amount of storage available from all local instance storage, in GB.

instance-storage-supported - Indicates whether the instance type has local instance storage (true | false).

instance-type - The instance type (for example c5.2xlarge or c5*).

memory-info.size-in-mib - The memory size.

network-info.efa-info.maximum-efa-interfaces - The maximum number of Elastic Fabric Adapters (EFAs) per instance.

network-info.efa-supported - Indicates whether the instance type supports Elastic Fabric Adapter (EFA) (true | false).

network-info.ena-support - Indicates whether Elastic Network Adapter (ENA) is supported or required (required | supported | unsupported).

network-info.encryption-in-transit-supported - Indicates whether the instance type automatically encrypts in-transit traffic between instances (true | false).

network-info.ipv4-addresses-per-interface - The maximum number of private IPv4 addresses per network interface.

network-info.ipv6-addresses-per-interface - The maximum number of private IPv6 addresses per network interface.

network-info.ipv6-supported - Indicates whether the instance type supports IPv6 (true | false).

network-info.maximum-network-interfaces - The maximum number of network interfaces per instance.

network-info.network-performance - The network performance (for example, "25 Gigabit").

processor-info.supported-architecture - The CPU architecture (arm64 | i386 | x86_64).

processor-info.sustained-clock-speed-in-ghz - The CPU clock speed, in GHz.

supported-boot-mode - The boot mode (legacy-bios | uefi).

supported-root-device-type - The root device type (ebs | instance-store).

supported-usage-class - The usage class (on-demand | spot).

supported-virtualization-type - The virtualization type (hvm | paravirtual).

vcpu-info.default-cores - The default number of cores for the instance type.

vcpu-info.default-threads-per-core - The default number of threads per core for the instance type.

vcpu-info.default-vcpus - The default number of vCPUs for the instance type.

vcpu-info.valid-cores - The number of cores that can be configured for the instance type.

vcpu-info.valid-threads-per-core - The number of threads per core that can be configured for the instance type. For example, "1" or "1,2".

$sel:instanceTypes:DescribeInstanceTypes', describeInstanceTypes_instanceTypes - The instance types. For more information, see Instance types in the Amazon EC2 User Guide.

Request Lenses

describeInstanceTypes_nextToken :: Lens' DescribeInstanceTypes (Maybe Text)

The token to retrieve the next page of results.

describeInstanceTypes_maxResults :: Lens' DescribeInstanceTypes (Maybe Natural)

The maximum number of results to return for the request in a single page. The remaining results can be seen by sending another request with the next token value.

describeInstanceTypes_dryRun :: Lens' DescribeInstanceTypes (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeInstanceTypes_filters :: Lens' DescribeInstanceTypes (Maybe [Filter])

One or more filters. Filter names and values are case-sensitive.

auto-recovery-supported - Indicates whether auto recovery is supported (true | false).

bare-metal - Indicates whether it is a bare metal instance type (true | false).

burstable-performance-supported - Indicates whether it is a burstable performance instance type (true | false).

current-generation - Indicates whether this instance type is the latest generation instance type of an instance family (true | false).

ebs-info.ebs-optimized-info.baseline-bandwidth-in-mbps - The baseline bandwidth performance for an EBS-optimized instance type, in Mbps.

ebs-info.ebs-optimized-info.baseline-iops - The baseline input/output storage operations per second for an EBS-optimized instance type.

ebs-info.ebs-optimized-info.baseline-throughput-in-mbps - The baseline throughput performance for an EBS-optimized instance type, in MB/s.

ebs-info.ebs-optimized-info.maximum-bandwidth-in-mbps - The maximum bandwidth performance for an EBS-optimized instance type, in Mbps.

ebs-info.ebs-optimized-info.maximum-iops - The maximum input/output storage operations per second for an EBS-optimized instance type.

ebs-info.ebs-optimized-info.maximum-throughput-in-mbps - The maximum throughput performance for an EBS-optimized instance type, in MB/s.

ebs-info.ebs-optimized-support - Indicates whether the instance type is EBS-optimized (supported | unsupported | default).

ebs-info.encryption-support - Indicates whether EBS encryption is supported (supported | unsupported).

ebs-info.nvme-support - Indicates whether non-volatile memory express (NVMe) is supported for EBS volumes (required | supported | unsupported).

free-tier-eligible - Indicates whether the instance type is eligible to use in the free tier (true | false).

hibernation-supported - Indicates whether On-Demand hibernation is supported (true | false).

hypervisor - The hypervisor (nitro | xen).

instance-storage-info.disk.count - The number of local disks.

instance-storage-info.disk.size-in-gb - The storage size of each instance storage disk, in GB.

instance-storage-info.disk.type - The storage technology for the local instance storage disks (hdd | ssd).

instance-storage-info.nvme-support - Indicates whether non-volatile memory express (NVMe) is supported for instance store (required | supported) | unsupported).

instance-storage-info.total-size-in-gb - The total amount of storage available from all local instance storage, in GB.

instance-storage-supported - Indicates whether the instance type has local instance storage (true | false).

instance-type - The instance type (for example c5.2xlarge or c5*).

memory-info.size-in-mib - The memory size.

network-info.efa-info.maximum-efa-interfaces - The maximum number of Elastic Fabric Adapters (EFAs) per instance.

network-info.efa-supported - Indicates whether the instance type supports Elastic Fabric Adapter (EFA) (true | false).

network-info.ena-support - Indicates whether Elastic Network Adapter (ENA) is supported or required (required | supported | unsupported).

network-info.encryption-in-transit-supported - Indicates whether the instance type automatically encrypts in-transit traffic between instances (true | false).

network-info.ipv4-addresses-per-interface - The maximum number of private IPv4 addresses per network interface.

network-info.ipv6-addresses-per-interface - The maximum number of private IPv6 addresses per network interface.

network-info.ipv6-supported - Indicates whether the instance type supports IPv6 (true | false).

network-info.maximum-network-interfaces - The maximum number of network interfaces per instance.

network-info.network-performance - The network performance (for example, "25 Gigabit").

processor-info.supported-architecture - The CPU architecture (arm64 | i386 | x86_64).

processor-info.sustained-clock-speed-in-ghz - The CPU clock speed, in GHz.

supported-boot-mode - The boot mode (legacy-bios | uefi).

supported-root-device-type - The root device type (ebs | instance-store).

supported-usage-class - The usage class (on-demand | spot).

supported-virtualization-type - The virtualization type (hvm | paravirtual).

vcpu-info.default-cores - The default number of cores for the instance type.

vcpu-info.default-threads-per-core - The default number of threads per core for the instance type.

vcpu-info.default-vcpus - The default number of vCPUs for the instance type.

vcpu-info.valid-cores - The number of cores that can be configured for the instance type.

vcpu-info.valid-threads-per-core - The number of threads per core that can be configured for the instance type. For example, "1" or "1,2".

describeInstanceTypes_instanceTypes :: Lens' DescribeInstanceTypes (Maybe [InstanceType])

The instance types. For more information, see Instance types in the Amazon EC2 User Guide.

Destructuring the Response

data DescribeInstanceTypesResponse

See: newDescribeInstanceTypesResponse smart constructor.

Constructors

DescribeInstanceTypesResponse'

Fields

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

instanceTypes :: Maybe [InstanceTypeInfo]

The instance type. For more information, see Instance types in the Amazon EC2 User Guide.

httpStatus :: Int

The response's http status code.

newDescribeInstanceTypesResponse

Arguments

:: Int
$sel:httpStatus:DescribeInstanceTypesResponse'

-> DescribeInstanceTypesResponse

Create a value of DescribeInstanceTypesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeInstanceTypes', describeInstanceTypesResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:instanceTypes:DescribeInstanceTypes', describeInstanceTypesResponse_instanceTypes - The instance type. For more information, see Instance types in the Amazon EC2 User Guide.

$sel:httpStatus:DescribeInstanceTypesResponse', describeInstanceTypesResponse_httpStatus - The response's http status code.

Response Lenses

describeInstanceTypesResponse_nextToken :: Lens' DescribeInstanceTypesResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

describeInstanceTypesResponse_instanceTypes :: Lens' DescribeInstanceTypesResponse (Maybe [InstanceTypeInfo])

The instance type. For more information, see Instance types in the Amazon EC2 User Guide.

describeInstanceTypesResponse_httpStatus :: Lens' DescribeInstanceTypesResponse Int

The response's http status code.