Network.AWS.EC2.DescribeInternetGateways

Description

Describes one or more of your internet gateways.

This operation returns paginated results.

Creating a Request

data DescribeInternetGateways

See: newDescribeInternetGateways smart constructor.

Constructors

DescribeInternetGateways'

Fields

nextToken :: Maybe Text

The token for the next page of results.

maxResults :: Maybe Natural

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

internetGatewayIds :: Maybe [Text]

One or more internet gateway IDs.

Default: Describes all your internet gateways.

filters :: Maybe [Filter]

One or more filters.

attachment.state - The current state of the attachment between the gateway and the VPC (available). Present only if a VPC is attached.

attachment.vpc-id - The ID of an attached VPC.

internet-gateway-id - The ID of the Internet gateway.

owner-id - The ID of the Amazon Web Services account that owns the internet gateway.

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

newDescribeInternetGateways :: DescribeInternetGateways

Create a value of DescribeInternetGateways with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeInternetGateways', describeInternetGateways_nextToken - The token for the next page of results.

$sel:maxResults:DescribeInternetGateways', describeInternetGateways_maxResults - The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

$sel:dryRun:DescribeInternetGateways', describeInternetGateways_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:internetGatewayIds:DescribeInternetGateways', describeInternetGateways_internetGatewayIds - One or more internet gateway IDs.

Default: Describes all your internet gateways.

$sel:filters:DescribeInternetGateways', describeInternetGateways_filters - One or more filters.

attachment.state - The current state of the attachment between the gateway and the VPC (available). Present only if a VPC is attached.

attachment.vpc-id - The ID of an attached VPC.

internet-gateway-id - The ID of the Internet gateway.

owner-id - The ID of the Amazon Web Services account that owns the internet gateway.

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

Request Lenses

describeInternetGateways_nextToken :: Lens' DescribeInternetGateways (Maybe Text)

The token for the next page of results.

describeInternetGateways_maxResults :: Lens' DescribeInternetGateways (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

describeInternetGateways_dryRun :: Lens' DescribeInternetGateways (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeInternetGateways_internetGatewayIds :: Lens' DescribeInternetGateways (Maybe [Text])

One or more internet gateway IDs.

Default: Describes all your internet gateways.

describeInternetGateways_filters :: Lens' DescribeInternetGateways (Maybe [Filter])

One or more filters.

attachment.state - The current state of the attachment between the gateway and the VPC (available). Present only if a VPC is attached.

attachment.vpc-id - The ID of an attached VPC.

internet-gateway-id - The ID of the Internet gateway.

owner-id - The ID of the Amazon Web Services account that owns the internet gateway.

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

Destructuring the Response

data DescribeInternetGatewaysResponse

See: newDescribeInternetGatewaysResponse smart constructor.

Constructors

DescribeInternetGatewaysResponse'

Fields

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

internetGateways :: Maybe [InternetGateway]

Information about one or more internet gateways.

httpStatus :: Int

The response's http status code.

newDescribeInternetGatewaysResponse

Arguments

:: Int
$sel:httpStatus:DescribeInternetGatewaysResponse'

-> DescribeInternetGatewaysResponse

Create a value of DescribeInternetGatewaysResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeInternetGateways', describeInternetGatewaysResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:internetGateways:DescribeInternetGatewaysResponse', describeInternetGatewaysResponse_internetGateways - Information about one or more internet gateways.

$sel:httpStatus:DescribeInternetGatewaysResponse', describeInternetGatewaysResponse_httpStatus - The response's http status code.

Response Lenses

describeInternetGatewaysResponse_nextToken :: Lens' DescribeInternetGatewaysResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

describeInternetGatewaysResponse_internetGateways :: Lens' DescribeInternetGatewaysResponse (Maybe [InternetGateway])

Information about one or more internet gateways.

describeInternetGatewaysResponse_httpStatus :: Lens' DescribeInternetGatewaysResponse Int

The response's http status code.