Network.AWS.EC2.DescribeLocalGateways

Description

Describes one or more local gateways. By default, all local gateways are described. Alternatively, you can filter the results.

This operation returns paginated results.

Creating a Request

data DescribeLocalGateways

See: newDescribeLocalGateways smart constructor.

Constructors

DescribeLocalGateways'

Fields

filters :: Maybe [Filter]

One or more filters.

nextToken :: Maybe Text

The token for the next page of results.

localGatewayIds :: Maybe [Text]

One or more filters.

local-gateway-id - The ID of a local gateway.

local-gateway-route-table-id - The ID of the local gateway route table.

local-gateway-route-table-virtual-interface-group-association-id - The ID of the association.

local-gateway-route-table-virtual-interface-group-id - The ID of the virtual interface group.

outpost-arn - The Amazon Resource Name (ARN) of the Outpost.

state - The state of the association.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

maxResults :: Maybe Natural

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

newDescribeLocalGateways :: DescribeLocalGateways

Create a value of DescribeLocalGateways with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:DescribeLocalGateways', describeLocalGateways_filters - One or more filters.

$sel:nextToken:DescribeLocalGateways', describeLocalGateways_nextToken - The token for the next page of results.

$sel:localGatewayIds:DescribeLocalGateways', describeLocalGateways_localGatewayIds - One or more filters.

local-gateway-id - The ID of a local gateway.

local-gateway-route-table-id - The ID of the local gateway route table.

local-gateway-route-table-virtual-interface-group-association-id - The ID of the association.

local-gateway-route-table-virtual-interface-group-id - The ID of the virtual interface group.

outpost-arn - The Amazon Resource Name (ARN) of the Outpost.

state - The state of the association.

$sel:dryRun:DescribeLocalGateways', describeLocalGateways_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:maxResults:DescribeLocalGateways', describeLocalGateways_maxResults - The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

Request Lenses

describeLocalGateways_filters :: Lens' DescribeLocalGateways (Maybe [Filter])

One or more filters.

describeLocalGateways_nextToken :: Lens' DescribeLocalGateways (Maybe Text)

The token for the next page of results.

describeLocalGateways_localGatewayIds :: Lens' DescribeLocalGateways (Maybe [Text])

One or more filters.

local-gateway-id - The ID of a local gateway.

local-gateway-route-table-id - The ID of the local gateway route table.

local-gateway-route-table-virtual-interface-group-association-id - The ID of the association.

local-gateway-route-table-virtual-interface-group-id - The ID of the virtual interface group.

outpost-arn - The Amazon Resource Name (ARN) of the Outpost.

state - The state of the association.

describeLocalGateways_dryRun :: Lens' DescribeLocalGateways (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeLocalGateways_maxResults :: Lens' DescribeLocalGateways (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

Destructuring the Response

data DescribeLocalGatewaysResponse

See: newDescribeLocalGatewaysResponse smart constructor.

Constructors

DescribeLocalGatewaysResponse'

Fields

localGateways :: Maybe [LocalGateway]

Information about the local gateways.

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

httpStatus :: Int

The response's http status code.

newDescribeLocalGatewaysResponse

Arguments

:: Int
$sel:httpStatus:DescribeLocalGatewaysResponse'

-> DescribeLocalGatewaysResponse

Create a value of DescribeLocalGatewaysResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:localGateways:DescribeLocalGatewaysResponse', describeLocalGatewaysResponse_localGateways - Information about the local gateways.

$sel:nextToken:DescribeLocalGateways', describeLocalGatewaysResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:httpStatus:DescribeLocalGatewaysResponse', describeLocalGatewaysResponse_httpStatus - The response's http status code.

Response Lenses

describeLocalGatewaysResponse_localGateways :: Lens' DescribeLocalGatewaysResponse (Maybe [LocalGateway])

Information about the local gateways.

describeLocalGatewaysResponse_nextToken :: Lens' DescribeLocalGatewaysResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

describeLocalGatewaysResponse_httpStatus :: Lens' DescribeLocalGatewaysResponse Int

The response's http status code.