Network.AWS.EC2.DescribeNatGateways

Description

Describes one or more of your NAT gateways.

This operation returns paginated results.

Creating a Request

data DescribeNatGateways

See: newDescribeNatGateways smart constructor.

Constructors

DescribeNatGateways'

Fields

nextToken :: Maybe Text

The token for the next page of results.

maxResults :: Maybe Natural

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

natGatewayIds :: Maybe [Text]

One or more NAT gateway IDs.

filter' :: Maybe [Filter]

One or more filters.

nat-gateway-id - The ID of the NAT gateway.

state - The state of the NAT gateway (pending | failed | available | deleting | deleted).

subnet-id - The ID of the subnet in which the NAT gateway resides.

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

vpc-id - The ID of the VPC in which the NAT gateway resides.

newDescribeNatGateways :: DescribeNatGateways

Create a value of DescribeNatGateways with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeNatGateways', describeNatGateways_nextToken - The token for the next page of results.

$sel:maxResults:DescribeNatGateways', describeNatGateways_maxResults - The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

$sel:dryRun:DescribeNatGateways', describeNatGateways_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:natGatewayIds:DescribeNatGateways', describeNatGateways_natGatewayIds - One or more NAT gateway IDs.

$sel:filter':DescribeNatGateways', describeNatGateways_filter - One or more filters.

nat-gateway-id - The ID of the NAT gateway.

state - The state of the NAT gateway (pending | failed | available | deleting | deleted).

subnet-id - The ID of the subnet in which the NAT gateway resides.

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

vpc-id - The ID of the VPC in which the NAT gateway resides.

Request Lenses

describeNatGateways_nextToken :: Lens' DescribeNatGateways (Maybe Text)

The token for the next page of results.

describeNatGateways_maxResults :: Lens' DescribeNatGateways (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

describeNatGateways_dryRun :: Lens' DescribeNatGateways (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeNatGateways_natGatewayIds :: Lens' DescribeNatGateways (Maybe [Text])

One or more NAT gateway IDs.

describeNatGateways_filter :: Lens' DescribeNatGateways (Maybe [Filter])

One or more filters.

nat-gateway-id - The ID of the NAT gateway.

state - The state of the NAT gateway (pending | failed | available | deleting | deleted).

subnet-id - The ID of the subnet in which the NAT gateway resides.

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

vpc-id - The ID of the VPC in which the NAT gateway resides.

Destructuring the Response

data DescribeNatGatewaysResponse

See: newDescribeNatGatewaysResponse smart constructor.

Constructors

DescribeNatGatewaysResponse'

Fields

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

natGateways :: Maybe [NatGateway]

Information about the NAT gateways.

httpStatus :: Int

The response's http status code.

newDescribeNatGatewaysResponse

Arguments

:: Int
$sel:httpStatus:DescribeNatGatewaysResponse'

-> DescribeNatGatewaysResponse

Create a value of DescribeNatGatewaysResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeNatGateways', describeNatGatewaysResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:natGateways:DescribeNatGatewaysResponse', describeNatGatewaysResponse_natGateways - Information about the NAT gateways.

$sel:httpStatus:DescribeNatGatewaysResponse', describeNatGatewaysResponse_httpStatus - The response's http status code.

Response Lenses

describeNatGatewaysResponse_nextToken :: Lens' DescribeNatGatewaysResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

describeNatGatewaysResponse_natGateways :: Lens' DescribeNatGatewaysResponse (Maybe [NatGateway])

Information about the NAT gateways.

describeNatGatewaysResponse_httpStatus :: Lens' DescribeNatGatewaysResponse Int

The response's http status code.