Network.AWS.EC2.DescribeNetworkInsightsAnalyses

Description

Describes one or more of your network insights analyses.

This operation returns paginated results.

Creating a Request

data DescribeNetworkInsightsAnalyses

See: newDescribeNetworkInsightsAnalyses smart constructor.

Constructors

DescribeNetworkInsightsAnalyses'

Fields

networkInsightsAnalysisIds :: Maybe [Text]

The ID of the network insights analyses. You must specify either analysis IDs or a path ID.

analysisStartTime :: Maybe ISO8601

The time when the network insights analyses started.

nextToken :: Maybe Text

The token for the next page of results.

maxResults :: Maybe Natural

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

networkInsightsPathId :: Maybe Text

The ID of the path. You must specify either a path ID or analysis IDs.

filters :: Maybe [Filter]

The filters. The following are possible values:

PathFound - A Boolean value that indicates whether a feasible path is found.

Status - The status of the analysis (running | succeeded | failed).

analysisEndTime :: Maybe ISO8601

The time when the network insights analyses ended.

newDescribeNetworkInsightsAnalyses :: DescribeNetworkInsightsAnalyses

Create a value of DescribeNetworkInsightsAnalyses with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:networkInsightsAnalysisIds:DescribeNetworkInsightsAnalyses', describeNetworkInsightsAnalyses_networkInsightsAnalysisIds - The ID of the network insights analyses. You must specify either analysis IDs or a path ID.

$sel:analysisStartTime:DescribeNetworkInsightsAnalyses', describeNetworkInsightsAnalyses_analysisStartTime - The time when the network insights analyses started.

$sel:nextToken:DescribeNetworkInsightsAnalyses', describeNetworkInsightsAnalyses_nextToken - The token for the next page of results.

$sel:maxResults:DescribeNetworkInsightsAnalyses', describeNetworkInsightsAnalyses_maxResults - The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

$sel:dryRun:DescribeNetworkInsightsAnalyses', describeNetworkInsightsAnalyses_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:networkInsightsPathId:DescribeNetworkInsightsAnalyses', describeNetworkInsightsAnalyses_networkInsightsPathId - The ID of the path. You must specify either a path ID or analysis IDs.

$sel:filters:DescribeNetworkInsightsAnalyses', describeNetworkInsightsAnalyses_filters - The filters. The following are possible values:

PathFound - A Boolean value that indicates whether a feasible path is found.

Status - The status of the analysis (running | succeeded | failed).

$sel:analysisEndTime:DescribeNetworkInsightsAnalyses', describeNetworkInsightsAnalyses_analysisEndTime - The time when the network insights analyses ended.

Request Lenses

describeNetworkInsightsAnalyses_networkInsightsAnalysisIds :: Lens' DescribeNetworkInsightsAnalyses (Maybe [Text])

The ID of the network insights analyses. You must specify either analysis IDs or a path ID.

describeNetworkInsightsAnalyses_analysisStartTime :: Lens' DescribeNetworkInsightsAnalyses (Maybe UTCTime)

The time when the network insights analyses started.

describeNetworkInsightsAnalyses_nextToken :: Lens' DescribeNetworkInsightsAnalyses (Maybe Text)

The token for the next page of results.

describeNetworkInsightsAnalyses_maxResults :: Lens' DescribeNetworkInsightsAnalyses (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

describeNetworkInsightsAnalyses_dryRun :: Lens' DescribeNetworkInsightsAnalyses (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeNetworkInsightsAnalyses_networkInsightsPathId :: Lens' DescribeNetworkInsightsAnalyses (Maybe Text)

The ID of the path. You must specify either a path ID or analysis IDs.

describeNetworkInsightsAnalyses_filters :: Lens' DescribeNetworkInsightsAnalyses (Maybe [Filter])

The filters. The following are possible values:

PathFound - A Boolean value that indicates whether a feasible path is found.

Status - The status of the analysis (running | succeeded | failed).

describeNetworkInsightsAnalyses_analysisEndTime :: Lens' DescribeNetworkInsightsAnalyses (Maybe UTCTime)

The time when the network insights analyses ended.

Destructuring the Response

data DescribeNetworkInsightsAnalysesResponse

See: newDescribeNetworkInsightsAnalysesResponse smart constructor.

Constructors

DescribeNetworkInsightsAnalysesResponse'

Fields

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

networkInsightsAnalyses :: Maybe [NetworkInsightsAnalysis]

Information about the network insights analyses.

httpStatus :: Int

The response's http status code.

newDescribeNetworkInsightsAnalysesResponse

Arguments

:: Int
$sel:httpStatus:DescribeNetworkInsightsAnalysesResponse'

-> DescribeNetworkInsightsAnalysesResponse

Create a value of DescribeNetworkInsightsAnalysesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeNetworkInsightsAnalyses', describeNetworkInsightsAnalysesResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:networkInsightsAnalyses:DescribeNetworkInsightsAnalysesResponse', describeNetworkInsightsAnalysesResponse_networkInsightsAnalyses - Information about the network insights analyses.

$sel:httpStatus:DescribeNetworkInsightsAnalysesResponse', describeNetworkInsightsAnalysesResponse_httpStatus - The response's http status code.

Response Lenses

describeNetworkInsightsAnalysesResponse_nextToken :: Lens' DescribeNetworkInsightsAnalysesResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

describeNetworkInsightsAnalysesResponse_networkInsightsAnalyses :: Lens' DescribeNetworkInsightsAnalysesResponse (Maybe [NetworkInsightsAnalysis])

Information about the network insights analyses.

describeNetworkInsightsAnalysesResponse_httpStatus :: Lens' DescribeNetworkInsightsAnalysesResponse Int

The response's http status code.