Network.AWS.EC2.DescribeNetworkInsightsPaths

Description

Describes one or more of your paths.

This operation returns paginated results.

Creating a Request

data DescribeNetworkInsightsPaths

See: newDescribeNetworkInsightsPaths smart constructor.

Constructors

DescribeNetworkInsightsPaths'

Fields

nextToken :: Maybe Text

The token for the next page of results.

maxResults :: Maybe Natural

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

networkInsightsPathIds :: Maybe [Text]

The IDs of the paths.

filters :: Maybe [Filter]

The filters. The following are possible values:

Destination - The ID of the resource.

DestinationPort - The destination port.

Name - The path name.

Protocol - The protocol.

Source - The ID of the resource.

newDescribeNetworkInsightsPaths :: DescribeNetworkInsightsPaths

Create a value of DescribeNetworkInsightsPaths with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeNetworkInsightsPaths', describeNetworkInsightsPaths_nextToken - The token for the next page of results.

$sel:maxResults:DescribeNetworkInsightsPaths', describeNetworkInsightsPaths_maxResults - The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

$sel:dryRun:DescribeNetworkInsightsPaths', describeNetworkInsightsPaths_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:networkInsightsPathIds:DescribeNetworkInsightsPaths', describeNetworkInsightsPaths_networkInsightsPathIds - The IDs of the paths.

$sel:filters:DescribeNetworkInsightsPaths', describeNetworkInsightsPaths_filters - The filters. The following are possible values:

Destination - The ID of the resource.

DestinationPort - The destination port.

Name - The path name.

Protocol - The protocol.

Source - The ID of the resource.

Request Lenses

describeNetworkInsightsPaths_nextToken :: Lens' DescribeNetworkInsightsPaths (Maybe Text)

The token for the next page of results.

describeNetworkInsightsPaths_maxResults :: Lens' DescribeNetworkInsightsPaths (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

describeNetworkInsightsPaths_dryRun :: Lens' DescribeNetworkInsightsPaths (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeNetworkInsightsPaths_networkInsightsPathIds :: Lens' DescribeNetworkInsightsPaths (Maybe [Text])

The IDs of the paths.

describeNetworkInsightsPaths_filters :: Lens' DescribeNetworkInsightsPaths (Maybe [Filter])

The filters. The following are possible values:

Destination - The ID of the resource.

DestinationPort - The destination port.

Name - The path name.

Protocol - The protocol.

Source - The ID of the resource.

Destructuring the Response

data DescribeNetworkInsightsPathsResponse

See: newDescribeNetworkInsightsPathsResponse smart constructor.

Constructors

DescribeNetworkInsightsPathsResponse'

Fields

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

networkInsightsPaths :: Maybe [NetworkInsightsPath]

Information about the paths.

httpStatus :: Int

The response's http status code.

newDescribeNetworkInsightsPathsResponse

Arguments

:: Int
$sel:httpStatus:DescribeNetworkInsightsPathsResponse'

-> DescribeNetworkInsightsPathsResponse

Create a value of DescribeNetworkInsightsPathsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeNetworkInsightsPaths', describeNetworkInsightsPathsResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:networkInsightsPaths:DescribeNetworkInsightsPathsResponse', describeNetworkInsightsPathsResponse_networkInsightsPaths - Information about the paths.

$sel:httpStatus:DescribeNetworkInsightsPathsResponse', describeNetworkInsightsPathsResponse_httpStatus - The response's http status code.

Response Lenses

describeNetworkInsightsPathsResponse_nextToken :: Lens' DescribeNetworkInsightsPathsResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

describeNetworkInsightsPathsResponse_networkInsightsPaths :: Lens' DescribeNetworkInsightsPathsResponse (Maybe [NetworkInsightsPath])

Information about the paths.

describeNetworkInsightsPathsResponse_httpStatus :: Lens' DescribeNetworkInsightsPathsResponse Int

The response's http status code.