Network.AWS.EC2.DescribeNetworkInterfacePermissions

Description

Describes the permissions for your network interfaces.

This operation returns paginated results.

Creating a Request

data DescribeNetworkInterfacePermissions

Contains the parameters for DescribeNetworkInterfacePermissions.

See: newDescribeNetworkInterfacePermissions smart constructor.

Constructors

DescribeNetworkInterfacePermissions'

Fields

nextToken :: Maybe Text

The token to request the next page of results.

maxResults :: Maybe Natural

The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned NextToken value. If this parameter is not specified, up to 50 results are returned by default.

networkInterfacePermissionIds :: Maybe [Text]

One or more network interface permission IDs.

filters :: Maybe [Filter]

One or more filters.

network-interface-permission.network-interface-permission-id - The ID of the permission.

network-interface-permission.network-interface-id - The ID of the network interface.

network-interface-permission.aws-account-id - The Amazon Web Services account ID.

network-interface-permission.aws-service - The Amazon Web Service.

network-interface-permission.permission - The type of permission (INSTANCE-ATTACH | EIP-ASSOCIATE).

newDescribeNetworkInterfacePermissions :: DescribeNetworkInterfacePermissions

Create a value of DescribeNetworkInterfacePermissions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeNetworkInterfacePermissions', describeNetworkInterfacePermissions_nextToken - The token to request the next page of results.

$sel:maxResults:DescribeNetworkInterfacePermissions', describeNetworkInterfacePermissions_maxResults - The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned NextToken value. If this parameter is not specified, up to 50 results are returned by default.

$sel:networkInterfacePermissionIds:DescribeNetworkInterfacePermissions', describeNetworkInterfacePermissions_networkInterfacePermissionIds - One or more network interface permission IDs.

$sel:filters:DescribeNetworkInterfacePermissions', describeNetworkInterfacePermissions_filters - One or more filters.

network-interface-permission.network-interface-permission-id - The ID of the permission.

network-interface-permission.network-interface-id - The ID of the network interface.

network-interface-permission.aws-account-id - The Amazon Web Services account ID.

network-interface-permission.aws-service - The Amazon Web Service.

network-interface-permission.permission - The type of permission (INSTANCE-ATTACH | EIP-ASSOCIATE).

Request Lenses

describeNetworkInterfacePermissions_nextToken :: Lens' DescribeNetworkInterfacePermissions (Maybe Text)

The token to request the next page of results.

describeNetworkInterfacePermissions_maxResults :: Lens' DescribeNetworkInterfacePermissions (Maybe Natural)

The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned NextToken value. If this parameter is not specified, up to 50 results are returned by default.

describeNetworkInterfacePermissions_networkInterfacePermissionIds :: Lens' DescribeNetworkInterfacePermissions (Maybe [Text])

One or more network interface permission IDs.

describeNetworkInterfacePermissions_filters :: Lens' DescribeNetworkInterfacePermissions (Maybe [Filter])

One or more filters.

network-interface-permission.network-interface-permission-id - The ID of the permission.

network-interface-permission.network-interface-id - The ID of the network interface.

network-interface-permission.aws-account-id - The Amazon Web Services account ID.

network-interface-permission.aws-service - The Amazon Web Service.

network-interface-permission.permission - The type of permission (INSTANCE-ATTACH | EIP-ASSOCIATE).

Destructuring the Response

data DescribeNetworkInterfacePermissionsResponse

Contains the output for DescribeNetworkInterfacePermissions.

See: newDescribeNetworkInterfacePermissionsResponse smart constructor.

Constructors

DescribeNetworkInterfacePermissionsResponse'

Fields

nextToken :: Maybe Text

The token to use to retrieve the next page of results.

networkInterfacePermissions :: Maybe [NetworkInterfacePermission]

The network interface permissions.

httpStatus :: Int

The response's http status code.

newDescribeNetworkInterfacePermissionsResponse

Arguments

:: Int
$sel:httpStatus:DescribeNetworkInterfacePermissionsResponse'

-> DescribeNetworkInterfacePermissionsResponse

Create a value of DescribeNetworkInterfacePermissionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeNetworkInterfacePermissions', describeNetworkInterfacePermissionsResponse_nextToken - The token to use to retrieve the next page of results.

$sel:networkInterfacePermissions:DescribeNetworkInterfacePermissionsResponse', describeNetworkInterfacePermissionsResponse_networkInterfacePermissions - The network interface permissions.

$sel:httpStatus:DescribeNetworkInterfacePermissionsResponse', describeNetworkInterfacePermissionsResponse_httpStatus - The response's http status code.

Response Lenses

describeNetworkInterfacePermissionsResponse_nextToken :: Lens' DescribeNetworkInterfacePermissionsResponse (Maybe Text)

The token to use to retrieve the next page of results.

describeNetworkInterfacePermissionsResponse_networkInterfacePermissions :: Lens' DescribeNetworkInterfacePermissionsResponse (Maybe [NetworkInterfacePermission])

The network interface permissions.

describeNetworkInterfacePermissionsResponse_httpStatus :: Lens' DescribeNetworkInterfacePermissionsResponse Int

The response's http status code.