Network.AWS.EC2.DescribeNetworkInterfaces

Description

Describes one or more of your network interfaces.

This operation returns paginated results.

Creating a Request

data DescribeNetworkInterfaces

Contains the parameters for DescribeNetworkInterfaces.

See: newDescribeNetworkInterfaces smart constructor.

Constructors

DescribeNetworkInterfaces'

Fields

nextToken :: Maybe Text

The token to retrieve the next page of results.

maxResults :: Maybe Natural

The maximum number of items to return for this request. The request returns a token that you can specify in a subsequent call to get the next set of results. You cannot specify this parameter and the network interface IDs parameter in the same request.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

networkInterfaceIds :: Maybe [Text]

One or more network interface IDs.

Default: Describes all your network interfaces.

filters :: Maybe [Filter]

One or more filters.

addresses.private-ip-address - The private IPv4 addresses associated with the network interface.

addresses.primary - Whether the private IPv4 address is the primary IP address associated with the network interface.

addresses.association.public-ip - The association ID returned when the network interface was associated with the Elastic IP address (IPv4).

addresses.association.owner-id - The owner ID of the addresses associated with the network interface.

association.association-id - The association ID returned when the network interface was associated with an IPv4 address.

association.allocation-id - The allocation ID returned when you allocated the Elastic IP address (IPv4) for your network interface.

association.ip-owner-id - The owner of the Elastic IP address (IPv4) associated with the network interface.

association.public-ip - The address of the Elastic IP address (IPv4) bound to the network interface.

association.public-dns-name - The public DNS name for the network interface (IPv4).

attachment.attachment-id - The ID of the interface attachment.

attachment.attach-time - The time that the network interface was attached to an instance.

attachment.delete-on-termination - Indicates whether the attachment is deleted when an instance is terminated.

attachment.device-index - The device index to which the network interface is attached.

attachment.instance-id - The ID of the instance to which the network interface is attached.

attachment.instance-owner-id - The owner ID of the instance to which the network interface is attached.

attachment.status - The status of the attachment (attaching | attached | detaching | detached).

availability-zone - The Availability Zone of the network interface.

description - The description of the network interface.

group-id - The ID of a security group associated with the network interface.

group-name - The name of a security group associated with the network interface.

ipv6-addresses.ipv6-address - An IPv6 address associated with the network interface.

mac-address - The MAC address of the network interface.

network-interface-id - The ID of the network interface.

owner-id - The Amazon Web Services account ID of the network interface owner.

private-ip-address - The private IPv4 address or addresses of the network interface.

private-dns-name - The private DNS name of the network interface (IPv4).

requester-id - The alias or Amazon Web Services account ID of the principal or service that created the network interface.

requester-managed - Indicates whether the network interface is being managed by an Amazon Web Service (for example, Amazon Web Services Management Console, Auto Scaling, and so on).

source-dest-check - Indicates whether the network interface performs source/destination checking. A value of true means checking is enabled, and false means checking is disabled. The value must be false for the network interface to perform network address translation (NAT) in your VPC.

status - The status of the network interface. If the network interface is not attached to an instance, the status is available; if a network interface is attached to an instance the status is in-use.

subnet-id - The ID of the subnet for the network interface.

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

vpc-id - The ID of the VPC for the network interface.

newDescribeNetworkInterfaces :: DescribeNetworkInterfaces

Create a value of DescribeNetworkInterfaces with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeNetworkInterfaces', describeNetworkInterfaces_nextToken - The token to retrieve the next page of results.

$sel:maxResults:DescribeNetworkInterfaces', describeNetworkInterfaces_maxResults - The maximum number of items to return for this request. The request returns a token that you can specify in a subsequent call to get the next set of results. You cannot specify this parameter and the network interface IDs parameter in the same request.

$sel:dryRun:DescribeNetworkInterfaces', describeNetworkInterfaces_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:networkInterfaceIds:DescribeNetworkInterfaces', describeNetworkInterfaces_networkInterfaceIds - One or more network interface IDs.

Default: Describes all your network interfaces.

$sel:filters:DescribeNetworkInterfaces', describeNetworkInterfaces_filters - One or more filters.

addresses.private-ip-address - The private IPv4 addresses associated with the network interface.

addresses.primary - Whether the private IPv4 address is the primary IP address associated with the network interface.

addresses.association.public-ip - The association ID returned when the network interface was associated with the Elastic IP address (IPv4).

addresses.association.owner-id - The owner ID of the addresses associated with the network interface.

association.association-id - The association ID returned when the network interface was associated with an IPv4 address.

association.allocation-id - The allocation ID returned when you allocated the Elastic IP address (IPv4) for your network interface.

association.ip-owner-id - The owner of the Elastic IP address (IPv4) associated with the network interface.

association.public-ip - The address of the Elastic IP address (IPv4) bound to the network interface.

association.public-dns-name - The public DNS name for the network interface (IPv4).

attachment.attachment-id - The ID of the interface attachment.

attachment.attach-time - The time that the network interface was attached to an instance.

attachment.delete-on-termination - Indicates whether the attachment is deleted when an instance is terminated.

attachment.device-index - The device index to which the network interface is attached.

attachment.instance-id - The ID of the instance to which the network interface is attached.

attachment.instance-owner-id - The owner ID of the instance to which the network interface is attached.

attachment.status - The status of the attachment (attaching | attached | detaching | detached).

availability-zone - The Availability Zone of the network interface.

description - The description of the network interface.

group-id - The ID of a security group associated with the network interface.

group-name - The name of a security group associated with the network interface.

ipv6-addresses.ipv6-address - An IPv6 address associated with the network interface.

mac-address - The MAC address of the network interface.

network-interface-id - The ID of the network interface.

owner-id - The Amazon Web Services account ID of the network interface owner.

private-ip-address - The private IPv4 address or addresses of the network interface.

private-dns-name - The private DNS name of the network interface (IPv4).

requester-id - The alias or Amazon Web Services account ID of the principal or service that created the network interface.

requester-managed - Indicates whether the network interface is being managed by an Amazon Web Service (for example, Amazon Web Services Management Console, Auto Scaling, and so on).

source-dest-check - Indicates whether the network interface performs source/destination checking. A value of true means checking is enabled, and false means checking is disabled. The value must be false for the network interface to perform network address translation (NAT) in your VPC.

status - The status of the network interface. If the network interface is not attached to an instance, the status is available; if a network interface is attached to an instance the status is in-use.

subnet-id - The ID of the subnet for the network interface.

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

vpc-id - The ID of the VPC for the network interface.

Request Lenses

describeNetworkInterfaces_nextToken :: Lens' DescribeNetworkInterfaces (Maybe Text)

The token to retrieve the next page of results.

describeNetworkInterfaces_maxResults :: Lens' DescribeNetworkInterfaces (Maybe Natural)

The maximum number of items to return for this request. The request returns a token that you can specify in a subsequent call to get the next set of results. You cannot specify this parameter and the network interface IDs parameter in the same request.

describeNetworkInterfaces_dryRun :: Lens' DescribeNetworkInterfaces (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeNetworkInterfaces_networkInterfaceIds :: Lens' DescribeNetworkInterfaces (Maybe [Text])

One or more network interface IDs.

Default: Describes all your network interfaces.

describeNetworkInterfaces_filters :: Lens' DescribeNetworkInterfaces (Maybe [Filter])

One or more filters.

addresses.private-ip-address - The private IPv4 addresses associated with the network interface.

addresses.primary - Whether the private IPv4 address is the primary IP address associated with the network interface.

addresses.association.public-ip - The association ID returned when the network interface was associated with the Elastic IP address (IPv4).

addresses.association.owner-id - The owner ID of the addresses associated with the network interface.

association.association-id - The association ID returned when the network interface was associated with an IPv4 address.

association.allocation-id - The allocation ID returned when you allocated the Elastic IP address (IPv4) for your network interface.

association.ip-owner-id - The owner of the Elastic IP address (IPv4) associated with the network interface.

association.public-ip - The address of the Elastic IP address (IPv4) bound to the network interface.

association.public-dns-name - The public DNS name for the network interface (IPv4).

attachment.attachment-id - The ID of the interface attachment.

attachment.attach-time - The time that the network interface was attached to an instance.

attachment.delete-on-termination - Indicates whether the attachment is deleted when an instance is terminated.

attachment.device-index - The device index to which the network interface is attached.

attachment.instance-id - The ID of the instance to which the network interface is attached.

attachment.instance-owner-id - The owner ID of the instance to which the network interface is attached.

attachment.status - The status of the attachment (attaching | attached | detaching | detached).

availability-zone - The Availability Zone of the network interface.

description - The description of the network interface.

group-id - The ID of a security group associated with the network interface.

group-name - The name of a security group associated with the network interface.

ipv6-addresses.ipv6-address - An IPv6 address associated with the network interface.

mac-address - The MAC address of the network interface.

network-interface-id - The ID of the network interface.

owner-id - The Amazon Web Services account ID of the network interface owner.

private-ip-address - The private IPv4 address or addresses of the network interface.

private-dns-name - The private DNS name of the network interface (IPv4).

requester-id - The alias or Amazon Web Services account ID of the principal or service that created the network interface.

requester-managed - Indicates whether the network interface is being managed by an Amazon Web Service (for example, Amazon Web Services Management Console, Auto Scaling, and so on).

source-dest-check - Indicates whether the network interface performs source/destination checking. A value of true means checking is enabled, and false means checking is disabled. The value must be false for the network interface to perform network address translation (NAT) in your VPC.

status - The status of the network interface. If the network interface is not attached to an instance, the status is available; if a network interface is attached to an instance the status is in-use.

subnet-id - The ID of the subnet for the network interface.

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

vpc-id - The ID of the VPC for the network interface.

Destructuring the Response

data DescribeNetworkInterfacesResponse

Contains the output of DescribeNetworkInterfaces.

See: newDescribeNetworkInterfacesResponse smart constructor.

Constructors

DescribeNetworkInterfacesResponse'

Fields

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

networkInterfaces :: Maybe [NetworkInterface]

Information about one or more network interfaces.

httpStatus :: Int

The response's http status code.

newDescribeNetworkInterfacesResponse

Arguments

:: Int
$sel:httpStatus:DescribeNetworkInterfacesResponse'

-> DescribeNetworkInterfacesResponse

Create a value of DescribeNetworkInterfacesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeNetworkInterfaces', describeNetworkInterfacesResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:networkInterfaces:DescribeNetworkInterfacesResponse', describeNetworkInterfacesResponse_networkInterfaces - Information about one or more network interfaces.

$sel:httpStatus:DescribeNetworkInterfacesResponse', describeNetworkInterfacesResponse_httpStatus - The response's http status code.

Response Lenses

describeNetworkInterfacesResponse_nextToken :: Lens' DescribeNetworkInterfacesResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

describeNetworkInterfacesResponse_networkInterfaces :: Lens' DescribeNetworkInterfacesResponse (Maybe [NetworkInterface])

Information about one or more network interfaces.

describeNetworkInterfacesResponse_httpStatus :: Lens' DescribeNetworkInterfacesResponse Int

The response's http status code.