Network.AWS.EC2.DescribeReplaceRootVolumeTasks

Description

Describes a root volume replacement task. For more information, see Replace a root volume in the Amazon Elastic Compute Cloud User Guide.

This operation returns paginated results.

Creating a Request

data DescribeReplaceRootVolumeTasks

See: newDescribeReplaceRootVolumeTasks smart constructor.

Constructors

DescribeReplaceRootVolumeTasks'

Fields

nextToken :: Maybe Text

The token for the next page of results.

replaceRootVolumeTaskIds :: Maybe [Text]

The ID of the root volume replacement task to view.

maxResults :: Maybe Natural

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

filters :: Maybe [Filter]

Filter to use:

instance-id - The ID of the instance for which the root volume replacement task was created.

newDescribeReplaceRootVolumeTasks :: DescribeReplaceRootVolumeTasks

Create a value of DescribeReplaceRootVolumeTasks with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeReplaceRootVolumeTasks', describeReplaceRootVolumeTasks_nextToken - The token for the next page of results.

$sel:replaceRootVolumeTaskIds:DescribeReplaceRootVolumeTasks', describeReplaceRootVolumeTasks_replaceRootVolumeTaskIds - The ID of the root volume replacement task to view.

$sel:maxResults:DescribeReplaceRootVolumeTasks', describeReplaceRootVolumeTasks_maxResults - The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

$sel:dryRun:DescribeReplaceRootVolumeTasks', describeReplaceRootVolumeTasks_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:filters:DescribeReplaceRootVolumeTasks', describeReplaceRootVolumeTasks_filters - Filter to use:

instance-id - The ID of the instance for which the root volume replacement task was created.

Request Lenses

describeReplaceRootVolumeTasks_nextToken :: Lens' DescribeReplaceRootVolumeTasks (Maybe Text)

The token for the next page of results.

describeReplaceRootVolumeTasks_replaceRootVolumeTaskIds :: Lens' DescribeReplaceRootVolumeTasks (Maybe [Text])

The ID of the root volume replacement task to view.

describeReplaceRootVolumeTasks_maxResults :: Lens' DescribeReplaceRootVolumeTasks (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

describeReplaceRootVolumeTasks_dryRun :: Lens' DescribeReplaceRootVolumeTasks (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeReplaceRootVolumeTasks_filters :: Lens' DescribeReplaceRootVolumeTasks (Maybe [Filter])

Filter to use:

instance-id - The ID of the instance for which the root volume replacement task was created.

Destructuring the Response

data DescribeReplaceRootVolumeTasksResponse

See: newDescribeReplaceRootVolumeTasksResponse smart constructor.

Constructors

DescribeReplaceRootVolumeTasksResponse'

Fields

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

replaceRootVolumeTasks :: Maybe [ReplaceRootVolumeTask]

Information about the root volume replacement task.

httpStatus :: Int

The response's http status code.

newDescribeReplaceRootVolumeTasksResponse

Arguments

:: Int
$sel:httpStatus:DescribeReplaceRootVolumeTasksResponse'

-> DescribeReplaceRootVolumeTasksResponse

Create a value of DescribeReplaceRootVolumeTasksResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeReplaceRootVolumeTasks', describeReplaceRootVolumeTasksResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:replaceRootVolumeTasks:DescribeReplaceRootVolumeTasksResponse', describeReplaceRootVolumeTasksResponse_replaceRootVolumeTasks - Information about the root volume replacement task.

$sel:httpStatus:DescribeReplaceRootVolumeTasksResponse', describeReplaceRootVolumeTasksResponse_httpStatus - The response's http status code.

Response Lenses

describeReplaceRootVolumeTasksResponse_nextToken :: Lens' DescribeReplaceRootVolumeTasksResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

describeReplaceRootVolumeTasksResponse_replaceRootVolumeTasks :: Lens' DescribeReplaceRootVolumeTasksResponse (Maybe [ReplaceRootVolumeTask])

Information about the root volume replacement task.

describeReplaceRootVolumeTasksResponse_httpStatus :: Lens' DescribeReplaceRootVolumeTasksResponse Int

The response's http status code.