Network.AWS.EC2.DescribeReservedInstances

Description

Describes one or more of the Reserved Instances that you purchased.

For more information about Reserved Instances, see Reserved Instances in the Amazon EC2 User Guide.

Creating a Request

data DescribeReservedInstances

Contains the parameters for DescribeReservedInstances.

See: newDescribeReservedInstances smart constructor.

Constructors

DescribeReservedInstances'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

offeringClass :: Maybe OfferingClassType

Describes whether the Reserved Instance is Standard or Convertible.

filters :: Maybe [Filter]

One or more filters.

availability-zone - The Availability Zone where the Reserved Instance can be used.

duration - The duration of the Reserved Instance (one year or three years), in seconds (31536000 | 94608000).

end - The time when the Reserved Instance expires (for example, 2015-08-07T11:54:42.000Z).

fixed-price - The purchase price of the Reserved Instance (for example, 9800.0).

instance-type - The instance type that is covered by the reservation.

scope - The scope of the Reserved Instance (Region or Availability Zone).

product-description - The Reserved Instance product platform description. Instances that include (Amazon VPC) in the product platform description will only be displayed to EC2-Classic account holders and are for use with Amazon VPC (Linux/UNIX | Linux/UNIX (Amazon VPC) | SUSE Linux | SUSE Linux (Amazon VPC) | Red Hat Enterprise Linux | Red Hat Enterprise Linux (Amazon VPC) | Red Hat Enterprise Linux with HA (Amazon VPC) | Windows | Windows (Amazon VPC) | Windows with SQL Server Standard | Windows with SQL Server Standard (Amazon VPC) | Windows with SQL Server Web | Windows with SQL Server Web (Amazon VPC) | Windows with SQL Server Enterprise | Windows with SQL Server Enterprise (Amazon VPC)).

reserved-instances-id - The ID of the Reserved Instance.

start - The time at which the Reserved Instance purchase request was placed (for example, 2014-08-07T11:54:42.000Z).

state - The state of the Reserved Instance (payment-pending | active | payment-failed | retired).

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

usage-price - The usage price of the Reserved Instance, per hour (for example, 0.84).

offeringType :: Maybe OfferingTypeValues

The Reserved Instance offering type. If you are using tools that predate the 2011-11-01 API version, you only have access to the Medium Utilization Reserved Instance offering type.

reservedInstancesIds :: Maybe [Text]

One or more Reserved Instance IDs.

Default: Describes all your Reserved Instances, or only those otherwise specified.

newDescribeReservedInstances :: DescribeReservedInstances

Create a value of DescribeReservedInstances with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:DescribeReservedInstances', describeReservedInstances_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:offeringClass:DescribeReservedInstances', describeReservedInstances_offeringClass - Describes whether the Reserved Instance is Standard or Convertible.

$sel:filters:DescribeReservedInstances', describeReservedInstances_filters - One or more filters.

availability-zone - The Availability Zone where the Reserved Instance can be used.

duration - The duration of the Reserved Instance (one year or three years), in seconds (31536000 | 94608000).

end - The time when the Reserved Instance expires (for example, 2015-08-07T11:54:42.000Z).

fixed-price - The purchase price of the Reserved Instance (for example, 9800.0).

instance-type - The instance type that is covered by the reservation.

scope - The scope of the Reserved Instance (Region or Availability Zone).

product-description - The Reserved Instance product platform description. Instances that include (Amazon VPC) in the product platform description will only be displayed to EC2-Classic account holders and are for use with Amazon VPC (Linux/UNIX | Linux/UNIX (Amazon VPC) | SUSE Linux | SUSE Linux (Amazon VPC) | Red Hat Enterprise Linux | Red Hat Enterprise Linux (Amazon VPC) | Red Hat Enterprise Linux with HA (Amazon VPC) | Windows | Windows (Amazon VPC) | Windows with SQL Server Standard | Windows with SQL Server Standard (Amazon VPC) | Windows with SQL Server Web | Windows with SQL Server Web (Amazon VPC) | Windows with SQL Server Enterprise | Windows with SQL Server Enterprise (Amazon VPC)).

reserved-instances-id - The ID of the Reserved Instance.

start - The time at which the Reserved Instance purchase request was placed (for example, 2014-08-07T11:54:42.000Z).

state - The state of the Reserved Instance (payment-pending | active | payment-failed | retired).

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

usage-price - The usage price of the Reserved Instance, per hour (for example, 0.84).

$sel:offeringType:DescribeReservedInstances', describeReservedInstances_offeringType - The Reserved Instance offering type. If you are using tools that predate the 2011-11-01 API version, you only have access to the Medium Utilization Reserved Instance offering type.

$sel:reservedInstancesIds:DescribeReservedInstances', describeReservedInstances_reservedInstancesIds - One or more Reserved Instance IDs.

Default: Describes all your Reserved Instances, or only those otherwise specified.

Request Lenses

describeReservedInstances_dryRun :: Lens' DescribeReservedInstances (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeReservedInstances_offeringClass :: Lens' DescribeReservedInstances (Maybe OfferingClassType)

Describes whether the Reserved Instance is Standard or Convertible.

describeReservedInstances_filters :: Lens' DescribeReservedInstances (Maybe [Filter])

One or more filters.

availability-zone - The Availability Zone where the Reserved Instance can be used.

duration - The duration of the Reserved Instance (one year or three years), in seconds (31536000 | 94608000).

end - The time when the Reserved Instance expires (for example, 2015-08-07T11:54:42.000Z).

fixed-price - The purchase price of the Reserved Instance (for example, 9800.0).

instance-type - The instance type that is covered by the reservation.

scope - The scope of the Reserved Instance (Region or Availability Zone).

product-description - The Reserved Instance product platform description. Instances that include (Amazon VPC) in the product platform description will only be displayed to EC2-Classic account holders and are for use with Amazon VPC (Linux/UNIX | Linux/UNIX (Amazon VPC) | SUSE Linux | SUSE Linux (Amazon VPC) | Red Hat Enterprise Linux | Red Hat Enterprise Linux (Amazon VPC) | Red Hat Enterprise Linux with HA (Amazon VPC) | Windows | Windows (Amazon VPC) | Windows with SQL Server Standard | Windows with SQL Server Standard (Amazon VPC) | Windows with SQL Server Web | Windows with SQL Server Web (Amazon VPC) | Windows with SQL Server Enterprise | Windows with SQL Server Enterprise (Amazon VPC)).

reserved-instances-id - The ID of the Reserved Instance.

start - The time at which the Reserved Instance purchase request was placed (for example, 2014-08-07T11:54:42.000Z).

state - The state of the Reserved Instance (payment-pending | active | payment-failed | retired).

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

usage-price - The usage price of the Reserved Instance, per hour (for example, 0.84).

describeReservedInstances_offeringType :: Lens' DescribeReservedInstances (Maybe OfferingTypeValues)

The Reserved Instance offering type. If you are using tools that predate the 2011-11-01 API version, you only have access to the Medium Utilization Reserved Instance offering type.

describeReservedInstances_reservedInstancesIds :: Lens' DescribeReservedInstances (Maybe [Text])

One or more Reserved Instance IDs.

Default: Describes all your Reserved Instances, or only those otherwise specified.

Destructuring the Response

data DescribeReservedInstancesResponse

Contains the output for DescribeReservedInstances.

See: newDescribeReservedInstancesResponse smart constructor.

Constructors

DescribeReservedInstancesResponse'

Fields

reservedInstances :: Maybe [ReservedInstances]

A list of Reserved Instances.

httpStatus :: Int

The response's http status code.

newDescribeReservedInstancesResponse

Arguments

:: Int
$sel:httpStatus:DescribeReservedInstancesResponse'

-> DescribeReservedInstancesResponse

Create a value of DescribeReservedInstancesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reservedInstances:DescribeReservedInstancesResponse', describeReservedInstancesResponse_reservedInstances - A list of Reserved Instances.

$sel:httpStatus:DescribeReservedInstancesResponse', describeReservedInstancesResponse_httpStatus - The response's http status code.

Response Lenses

describeReservedInstancesResponse_reservedInstances :: Lens' DescribeReservedInstancesResponse (Maybe [ReservedInstances])

A list of Reserved Instances.

describeReservedInstancesResponse_httpStatus :: Lens' DescribeReservedInstancesResponse Int

The response's http status code.