Network.AWS.EC2.DescribeReservedInstancesListings

Description

Describes your account's Reserved Instance listings in the Reserved Instance Marketplace.

The Reserved Instance Marketplace matches sellers who want to resell Reserved Instance capacity that they no longer need with buyers who want to purchase additional capacity. Reserved Instances bought and sold through the Reserved Instance Marketplace work like any other Reserved Instances.

As a seller, you choose to list some or all of your Reserved Instances, and you specify the upfront price to receive for them. Your Reserved Instances are then listed in the Reserved Instance Marketplace and are available for purchase.

As a buyer, you specify the configuration of the Reserved Instance to purchase, and the Marketplace matches what you're searching for with what's available. The Marketplace first sells the lowest priced Reserved Instances to you, and continues to sell available Reserved Instance listings to you until your demand is met. You are charged based on the total price of all of the listings that you purchase.

For more information, see Reserved Instance Marketplace in the Amazon EC2 User Guide.

Creating a Request

data DescribeReservedInstancesListings

Contains the parameters for DescribeReservedInstancesListings.

See: newDescribeReservedInstancesListings smart constructor.

Constructors

DescribeReservedInstancesListings'

Fields

reservedInstancesId :: Maybe Text

One or more Reserved Instance IDs.

reservedInstancesListingId :: Maybe Text

One or more Reserved Instance listing IDs.

filters :: Maybe [Filter]

One or more filters.

reserved-instances-id - The ID of the Reserved Instances.

reserved-instances-listing-id - The ID of the Reserved Instances listing.

status - The status of the Reserved Instance listing (pending | active | cancelled | closed).

status-message - The reason for the status.

newDescribeReservedInstancesListings :: DescribeReservedInstancesListings

Create a value of DescribeReservedInstancesListings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reservedInstancesId:DescribeReservedInstancesListings', describeReservedInstancesListings_reservedInstancesId - One or more Reserved Instance IDs.

$sel:reservedInstancesListingId:DescribeReservedInstancesListings', describeReservedInstancesListings_reservedInstancesListingId - One or more Reserved Instance listing IDs.

$sel:filters:DescribeReservedInstancesListings', describeReservedInstancesListings_filters - One or more filters.

reserved-instances-id - The ID of the Reserved Instances.

reserved-instances-listing-id - The ID of the Reserved Instances listing.

status - The status of the Reserved Instance listing (pending | active | cancelled | closed).

status-message - The reason for the status.

Request Lenses

describeReservedInstancesListings_reservedInstancesId :: Lens' DescribeReservedInstancesListings (Maybe Text)

One or more Reserved Instance IDs.

describeReservedInstancesListings_reservedInstancesListingId :: Lens' DescribeReservedInstancesListings (Maybe Text)

One or more Reserved Instance listing IDs.

describeReservedInstancesListings_filters :: Lens' DescribeReservedInstancesListings (Maybe [Filter])

One or more filters.

reserved-instances-id - The ID of the Reserved Instances.

reserved-instances-listing-id - The ID of the Reserved Instances listing.

status - The status of the Reserved Instance listing (pending | active | cancelled | closed).

status-message - The reason for the status.

Destructuring the Response

data DescribeReservedInstancesListingsResponse

Contains the output of DescribeReservedInstancesListings.

See: newDescribeReservedInstancesListingsResponse smart constructor.

Constructors

DescribeReservedInstancesListingsResponse'

Fields

reservedInstancesListings :: Maybe [ReservedInstancesListing]

Information about the Reserved Instance listing.

httpStatus :: Int

The response's http status code.

newDescribeReservedInstancesListingsResponse

Arguments

:: Int
$sel:httpStatus:DescribeReservedInstancesListingsResponse'

-> DescribeReservedInstancesListingsResponse

Create a value of DescribeReservedInstancesListingsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reservedInstancesListings:DescribeReservedInstancesListingsResponse', describeReservedInstancesListingsResponse_reservedInstancesListings - Information about the Reserved Instance listing.

$sel:httpStatus:DescribeReservedInstancesListingsResponse', describeReservedInstancesListingsResponse_httpStatus - The response's http status code.

Response Lenses

describeReservedInstancesListingsResponse_reservedInstancesListings :: Lens' DescribeReservedInstancesListingsResponse (Maybe [ReservedInstancesListing])

Information about the Reserved Instance listing.

describeReservedInstancesListingsResponse_httpStatus :: Lens' DescribeReservedInstancesListingsResponse Int

The response's http status code.