Network.AWS.EC2.DescribeReservedInstancesOfferings

Description

Describes Reserved Instance offerings that are available for purchase. With Reserved Instances, you purchase the right to launch instances for a period of time. During that time period, you do not receive insufficient capacity errors, and you pay a lower usage rate than the rate charged for On-Demand instances for the actual time used.

If you have listed your own Reserved Instances for sale in the Reserved Instance Marketplace, they will be excluded from these results. This is to ensure that you do not purchase your own Reserved Instances.

For more information, see Reserved Instance Marketplace in the Amazon EC2 User Guide.

This operation returns paginated results.

Creating a Request

data DescribeReservedInstancesOfferings

Contains the parameters for DescribeReservedInstancesOfferings.

See: newDescribeReservedInstancesOfferings smart constructor.

Constructors

DescribeReservedInstancesOfferings'

Fields

maxDuration :: Maybe Integer

The maximum duration (in seconds) to filter when searching for offerings.

Default: 94608000 (3 years)

productDescription :: Maybe RIProductDescription

The Reserved Instance product platform description. Instances that include (Amazon VPC) in the description are for use with Amazon VPC.

filters :: Maybe [Filter]

One or more filters.

availability-zone - The Availability Zone where the Reserved Instance can be used.

duration - The duration of the Reserved Instance (for example, one year or three years), in seconds (31536000 | 94608000).

fixed-price - The purchase price of the Reserved Instance (for example, 9800.0).

instance-type - The instance type that is covered by the reservation.

marketplace - Set to true to show only Reserved Instance Marketplace offerings. When this filter is not used, which is the default behavior, all offerings from both Amazon Web Services and the Reserved Instance Marketplace are listed.

product-description - The Reserved Instance product platform description. Instances that include (Amazon VPC) in the product platform description will only be displayed to EC2-Classic account holders and are for use with Amazon VPC. (Linux/UNIX | Linux/UNIX (Amazon VPC) | SUSE Linux | SUSE Linux (Amazon VPC) | Red Hat Enterprise Linux | Red Hat Enterprise Linux (Amazon VPC) | Red Hat Enterprise Linux with HA (Amazon VPC) | Windows | Windows (Amazon VPC) | Windows with SQL Server Standard | Windows with SQL Server Standard (Amazon VPC) | Windows with SQL Server Web | Windows with SQL Server Web (Amazon VPC) | Windows with SQL Server Enterprise | Windows with SQL Server Enterprise (Amazon VPC))

reserved-instances-offering-id - The Reserved Instances offering ID.

scope - The scope of the Reserved Instance (Availability Zone or Region).

usage-price - The usage price of the Reserved Instance, per hour (for example, 0.84).

includeMarketplace :: Maybe Bool

Include Reserved Instance Marketplace offerings in the response.

instanceType :: Maybe InstanceType

The instance type that the reservation will cover (for example, m1.small). For more information, see Instance types in the Amazon EC2 User Guide.

nextToken :: Maybe Text

The token to retrieve the next page of results.

minDuration :: Maybe Integer

The minimum duration (in seconds) to filter when searching for offerings.

Default: 2592000 (1 month)

availabilityZone :: Maybe Text

The Availability Zone in which the Reserved Instance can be used.

offeringType :: Maybe OfferingTypeValues

The Reserved Instance offering type. If you are using tools that predate the 2011-11-01 API version, you only have access to the Medium Utilization Reserved Instance offering type.

reservedInstancesOfferingIds :: Maybe [Text]

One or more Reserved Instances offering IDs.

instanceTenancy :: Maybe Tenancy

The tenancy of the instances covered by the reservation. A Reserved Instance with a tenancy of dedicated is applied to instances that run in a VPC on single-tenant hardware (i.e., Dedicated Instances).

Important: The host value cannot be used with this parameter. Use the default or dedicated values only.

Default: default

offeringClass :: Maybe OfferingClassType

The offering class of the Reserved Instance. Can be standard or convertible.

maxInstanceCount :: Maybe Int

The maximum number of instances to filter when searching for offerings.

Default: 20

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

maxResults :: Maybe Int

The maximum number of results to return for the request in a single page. The remaining results of the initial request can be seen by sending another request with the returned NextToken value. The maximum is 100.

Default: 100

newDescribeReservedInstancesOfferings :: DescribeReservedInstancesOfferings

Create a value of DescribeReservedInstancesOfferings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxDuration:DescribeReservedInstancesOfferings', describeReservedInstancesOfferings_maxDuration - The maximum duration (in seconds) to filter when searching for offerings.

Default: 94608000 (3 years)

$sel:productDescription:DescribeReservedInstancesOfferings', describeReservedInstancesOfferings_productDescription - The Reserved Instance product platform description. Instances that include (Amazon VPC) in the description are for use with Amazon VPC.

$sel:filters:DescribeReservedInstancesOfferings', describeReservedInstancesOfferings_filters - One or more filters.

availability-zone - The Availability Zone where the Reserved Instance can be used.

duration - The duration of the Reserved Instance (for example, one year or three years), in seconds (31536000 | 94608000).

fixed-price - The purchase price of the Reserved Instance (for example, 9800.0).

instance-type - The instance type that is covered by the reservation.

marketplace - Set to true to show only Reserved Instance Marketplace offerings. When this filter is not used, which is the default behavior, all offerings from both Amazon Web Services and the Reserved Instance Marketplace are listed.

product-description - The Reserved Instance product platform description. Instances that include (Amazon VPC) in the product platform description will only be displayed to EC2-Classic account holders and are for use with Amazon VPC. (Linux/UNIX | Linux/UNIX (Amazon VPC) | SUSE Linux | SUSE Linux (Amazon VPC) | Red Hat Enterprise Linux | Red Hat Enterprise Linux (Amazon VPC) | Red Hat Enterprise Linux with HA (Amazon VPC) | Windows | Windows (Amazon VPC) | Windows with SQL Server Standard | Windows with SQL Server Standard (Amazon VPC) | Windows with SQL Server Web | Windows with SQL Server Web (Amazon VPC) | Windows with SQL Server Enterprise | Windows with SQL Server Enterprise (Amazon VPC))

reserved-instances-offering-id - The Reserved Instances offering ID.

scope - The scope of the Reserved Instance (Availability Zone or Region).

usage-price - The usage price of the Reserved Instance, per hour (for example, 0.84).

$sel:includeMarketplace:DescribeReservedInstancesOfferings', describeReservedInstancesOfferings_includeMarketplace - Include Reserved Instance Marketplace offerings in the response.

$sel:instanceType:DescribeReservedInstancesOfferings', describeReservedInstancesOfferings_instanceType - The instance type that the reservation will cover (for example, m1.small). For more information, see Instance types in the Amazon EC2 User Guide.

$sel:nextToken:DescribeReservedInstancesOfferings', describeReservedInstancesOfferings_nextToken - The token to retrieve the next page of results.

$sel:minDuration:DescribeReservedInstancesOfferings', describeReservedInstancesOfferings_minDuration - The minimum duration (in seconds) to filter when searching for offerings.

Default: 2592000 (1 month)

$sel:availabilityZone:DescribeReservedInstancesOfferings', describeReservedInstancesOfferings_availabilityZone - The Availability Zone in which the Reserved Instance can be used.

$sel:offeringType:DescribeReservedInstancesOfferings', describeReservedInstancesOfferings_offeringType - The Reserved Instance offering type. If you are using tools that predate the 2011-11-01 API version, you only have access to the Medium Utilization Reserved Instance offering type.

$sel:reservedInstancesOfferingIds:DescribeReservedInstancesOfferings', describeReservedInstancesOfferings_reservedInstancesOfferingIds - One or more Reserved Instances offering IDs.

$sel:instanceTenancy:DescribeReservedInstancesOfferings', describeReservedInstancesOfferings_instanceTenancy - The tenancy of the instances covered by the reservation. A Reserved Instance with a tenancy of dedicated is applied to instances that run in a VPC on single-tenant hardware (i.e., Dedicated Instances).

Important: The host value cannot be used with this parameter. Use the default or dedicated values only.

Default: default

$sel:offeringClass:DescribeReservedInstancesOfferings', describeReservedInstancesOfferings_offeringClass - The offering class of the Reserved Instance. Can be standard or convertible.

$sel:maxInstanceCount:DescribeReservedInstancesOfferings', describeReservedInstancesOfferings_maxInstanceCount - The maximum number of instances to filter when searching for offerings.

Default: 20

$sel:dryRun:DescribeReservedInstancesOfferings', describeReservedInstancesOfferings_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:maxResults:DescribeReservedInstancesOfferings', describeReservedInstancesOfferings_maxResults - The maximum number of results to return for the request in a single page. The remaining results of the initial request can be seen by sending another request with the returned NextToken value. The maximum is 100.

Default: 100

Request Lenses

describeReservedInstancesOfferings_maxDuration :: Lens' DescribeReservedInstancesOfferings (Maybe Integer)

The maximum duration (in seconds) to filter when searching for offerings.

Default: 94608000 (3 years)

describeReservedInstancesOfferings_productDescription :: Lens' DescribeReservedInstancesOfferings (Maybe RIProductDescription)

The Reserved Instance product platform description. Instances that include (Amazon VPC) in the description are for use with Amazon VPC.

describeReservedInstancesOfferings_filters :: Lens' DescribeReservedInstancesOfferings (Maybe [Filter])

One or more filters.

availability-zone - The Availability Zone where the Reserved Instance can be used.

duration - The duration of the Reserved Instance (for example, one year or three years), in seconds (31536000 | 94608000).

fixed-price - The purchase price of the Reserved Instance (for example, 9800.0).

instance-type - The instance type that is covered by the reservation.

marketplace - Set to true to show only Reserved Instance Marketplace offerings. When this filter is not used, which is the default behavior, all offerings from both Amazon Web Services and the Reserved Instance Marketplace are listed.

product-description - The Reserved Instance product platform description. Instances that include (Amazon VPC) in the product platform description will only be displayed to EC2-Classic account holders and are for use with Amazon VPC. (Linux/UNIX | Linux/UNIX (Amazon VPC) | SUSE Linux | SUSE Linux (Amazon VPC) | Red Hat Enterprise Linux | Red Hat Enterprise Linux (Amazon VPC) | Red Hat Enterprise Linux with HA (Amazon VPC) | Windows | Windows (Amazon VPC) | Windows with SQL Server Standard | Windows with SQL Server Standard (Amazon VPC) | Windows with SQL Server Web | Windows with SQL Server Web (Amazon VPC) | Windows with SQL Server Enterprise | Windows with SQL Server Enterprise (Amazon VPC))

reserved-instances-offering-id - The Reserved Instances offering ID.

scope - The scope of the Reserved Instance (Availability Zone or Region).

usage-price - The usage price of the Reserved Instance, per hour (for example, 0.84).

describeReservedInstancesOfferings_includeMarketplace :: Lens' DescribeReservedInstancesOfferings (Maybe Bool)

Include Reserved Instance Marketplace offerings in the response.

describeReservedInstancesOfferings_instanceType :: Lens' DescribeReservedInstancesOfferings (Maybe InstanceType)

The instance type that the reservation will cover (for example, m1.small). For more information, see Instance types in the Amazon EC2 User Guide.

describeReservedInstancesOfferings_nextToken :: Lens' DescribeReservedInstancesOfferings (Maybe Text)

The token to retrieve the next page of results.

describeReservedInstancesOfferings_minDuration :: Lens' DescribeReservedInstancesOfferings (Maybe Integer)

The minimum duration (in seconds) to filter when searching for offerings.

Default: 2592000 (1 month)

describeReservedInstancesOfferings_availabilityZone :: Lens' DescribeReservedInstancesOfferings (Maybe Text)

The Availability Zone in which the Reserved Instance can be used.

describeReservedInstancesOfferings_offeringType :: Lens' DescribeReservedInstancesOfferings (Maybe OfferingTypeValues)

The Reserved Instance offering type. If you are using tools that predate the 2011-11-01 API version, you only have access to the Medium Utilization Reserved Instance offering type.

describeReservedInstancesOfferings_reservedInstancesOfferingIds :: Lens' DescribeReservedInstancesOfferings (Maybe [Text])

One or more Reserved Instances offering IDs.

describeReservedInstancesOfferings_instanceTenancy :: Lens' DescribeReservedInstancesOfferings (Maybe Tenancy)

The tenancy of the instances covered by the reservation. A Reserved Instance with a tenancy of dedicated is applied to instances that run in a VPC on single-tenant hardware (i.e., Dedicated Instances).

Important: The host value cannot be used with this parameter. Use the default or dedicated values only.

Default: default

describeReservedInstancesOfferings_offeringClass :: Lens' DescribeReservedInstancesOfferings (Maybe OfferingClassType)

The offering class of the Reserved Instance. Can be standard or convertible.

describeReservedInstancesOfferings_maxInstanceCount :: Lens' DescribeReservedInstancesOfferings (Maybe Int)

The maximum number of instances to filter when searching for offerings.

Default: 20

describeReservedInstancesOfferings_dryRun :: Lens' DescribeReservedInstancesOfferings (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeReservedInstancesOfferings_maxResults :: Lens' DescribeReservedInstancesOfferings (Maybe Int)

The maximum number of results to return for the request in a single page. The remaining results of the initial request can be seen by sending another request with the returned NextToken value. The maximum is 100.

Default: 100

Destructuring the Response

data DescribeReservedInstancesOfferingsResponse

Contains the output of DescribeReservedInstancesOfferings.

See: newDescribeReservedInstancesOfferingsResponse smart constructor.

Constructors

DescribeReservedInstancesOfferingsResponse'

Fields

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

reservedInstancesOfferings :: Maybe [ReservedInstancesOffering]

A list of Reserved Instances offerings.

httpStatus :: Int

The response's http status code.

newDescribeReservedInstancesOfferingsResponse

Arguments

:: Int
$sel:httpStatus:DescribeReservedInstancesOfferingsResponse'

-> DescribeReservedInstancesOfferingsResponse

Create a value of DescribeReservedInstancesOfferingsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeReservedInstancesOfferings', describeReservedInstancesOfferingsResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:reservedInstancesOfferings:DescribeReservedInstancesOfferingsResponse', describeReservedInstancesOfferingsResponse_reservedInstancesOfferings - A list of Reserved Instances offerings.

$sel:httpStatus:DescribeReservedInstancesOfferingsResponse', describeReservedInstancesOfferingsResponse_httpStatus - The response's http status code.

Response Lenses

describeReservedInstancesOfferingsResponse_nextToken :: Lens' DescribeReservedInstancesOfferingsResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

describeReservedInstancesOfferingsResponse_reservedInstancesOfferings :: Lens' DescribeReservedInstancesOfferingsResponse (Maybe [ReservedInstancesOffering])

A list of Reserved Instances offerings.

describeReservedInstancesOfferingsResponse_httpStatus :: Lens' DescribeReservedInstancesOfferingsResponse Int

The response's http status code.