Network.AWS.EC2.DescribeScheduledInstanceAvailability

Description

Finds available schedules that meet the specified criteria.

You can search for an available schedule no more than 3 months in advance. You must meet the minimum required duration of 1,200 hours per year. For example, the minimum daily schedule is 4 hours, the minimum weekly schedule is 24 hours, and the minimum monthly schedule is 100 hours.

After you find a schedule that meets your needs, call PurchaseScheduledInstances to purchase Scheduled Instances with that schedule.

This operation returns paginated results.

Creating a Request

data DescribeScheduledInstanceAvailability

Contains the parameters for DescribeScheduledInstanceAvailability.

See: newDescribeScheduledInstanceAvailability smart constructor.

Constructors

DescribeScheduledInstanceAvailability'

Fields

minSlotDurationInHours :: Maybe Int

The minimum available duration, in hours. The minimum required duration is 1,200 hours per year. For example, the minimum daily schedule is 4 hours, the minimum weekly schedule is 24 hours, and the minimum monthly schedule is 100 hours.

filters :: Maybe [Filter]

The filters.

availability-zone - The Availability Zone (for example, us-west-2a).

instance-type - The instance type (for example, c4.large).

network-platform - The network platform (EC2-Classic or EC2-VPC).

platform - The platform (Linux/UNIX or Windows).

nextToken :: Maybe Text

The token for the next set of results.

maxSlotDurationInHours :: Maybe Int

The maximum available duration, in hours. This value must be greater than MinSlotDurationInHours and less than 1,720.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

maxResults :: Maybe Natural

The maximum number of results to return in a single call. This value can be between 5 and 300. The default value is 300. To retrieve the remaining results, make another call with the returned NextToken value.

firstSlotStartTimeRange :: SlotDateTimeRangeRequest

The time period for the first schedule to start.

recurrence :: ScheduledInstanceRecurrenceRequest

The schedule recurrence.

newDescribeScheduledInstanceAvailability

Arguments

:: SlotDateTimeRangeRequest
$sel:firstSlotStartTimeRange:DescribeScheduledInstanceAvailability'

-> ScheduledInstanceRecurrenceRequest
$sel:recurrence:DescribeScheduledInstanceAvailability'

-> DescribeScheduledInstanceAvailability

Create a value of DescribeScheduledInstanceAvailability with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:minSlotDurationInHours:DescribeScheduledInstanceAvailability', describeScheduledInstanceAvailability_minSlotDurationInHours - The minimum available duration, in hours. The minimum required duration is 1,200 hours per year. For example, the minimum daily schedule is 4 hours, the minimum weekly schedule is 24 hours, and the minimum monthly schedule is 100 hours.

$sel:filters:DescribeScheduledInstanceAvailability', describeScheduledInstanceAvailability_filters - The filters.

availability-zone - The Availability Zone (for example, us-west-2a).

instance-type - The instance type (for example, c4.large).

network-platform - The network platform (EC2-Classic or EC2-VPC).

platform - The platform (Linux/UNIX or Windows).

$sel:nextToken:DescribeScheduledInstanceAvailability', describeScheduledInstanceAvailability_nextToken - The token for the next set of results.

$sel:maxSlotDurationInHours:DescribeScheduledInstanceAvailability', describeScheduledInstanceAvailability_maxSlotDurationInHours - The maximum available duration, in hours. This value must be greater than MinSlotDurationInHours and less than 1,720.

$sel:dryRun:DescribeScheduledInstanceAvailability', describeScheduledInstanceAvailability_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:maxResults:DescribeScheduledInstanceAvailability', describeScheduledInstanceAvailability_maxResults - The maximum number of results to return in a single call. This value can be between 5 and 300. The default value is 300. To retrieve the remaining results, make another call with the returned NextToken value.

$sel:firstSlotStartTimeRange:DescribeScheduledInstanceAvailability', describeScheduledInstanceAvailability_firstSlotStartTimeRange - The time period for the first schedule to start.

$sel:recurrence:DescribeScheduledInstanceAvailability', describeScheduledInstanceAvailability_recurrence - The schedule recurrence.

Request Lenses

describeScheduledInstanceAvailability_minSlotDurationInHours :: Lens' DescribeScheduledInstanceAvailability (Maybe Int)

The minimum available duration, in hours. The minimum required duration is 1,200 hours per year. For example, the minimum daily schedule is 4 hours, the minimum weekly schedule is 24 hours, and the minimum monthly schedule is 100 hours.

describeScheduledInstanceAvailability_filters :: Lens' DescribeScheduledInstanceAvailability (Maybe [Filter])

The filters.

availability-zone - The Availability Zone (for example, us-west-2a).

instance-type - The instance type (for example, c4.large).

network-platform - The network platform (EC2-Classic or EC2-VPC).

platform - The platform (Linux/UNIX or Windows).

describeScheduledInstanceAvailability_nextToken :: Lens' DescribeScheduledInstanceAvailability (Maybe Text)

The token for the next set of results.

describeScheduledInstanceAvailability_maxSlotDurationInHours :: Lens' DescribeScheduledInstanceAvailability (Maybe Int)

The maximum available duration, in hours. This value must be greater than MinSlotDurationInHours and less than 1,720.

describeScheduledInstanceAvailability_dryRun :: Lens' DescribeScheduledInstanceAvailability (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeScheduledInstanceAvailability_maxResults :: Lens' DescribeScheduledInstanceAvailability (Maybe Natural)

The maximum number of results to return in a single call. This value can be between 5 and 300. The default value is 300. To retrieve the remaining results, make another call with the returned NextToken value.

describeScheduledInstanceAvailability_firstSlotStartTimeRange :: Lens' DescribeScheduledInstanceAvailability SlotDateTimeRangeRequest

The time period for the first schedule to start.

describeScheduledInstanceAvailability_recurrence :: Lens' DescribeScheduledInstanceAvailability ScheduledInstanceRecurrenceRequest

The schedule recurrence.

Destructuring the Response

data DescribeScheduledInstanceAvailabilityResponse

Contains the output of DescribeScheduledInstanceAvailability.

See: newDescribeScheduledInstanceAvailabilityResponse smart constructor.

Constructors

DescribeScheduledInstanceAvailabilityResponse'

Fields

scheduledInstanceAvailabilitySet :: Maybe [ScheduledInstanceAvailability]

Information about the available Scheduled Instances.

nextToken :: Maybe Text

The token required to retrieve the next set of results. This value is null when there are no more results to return.

httpStatus :: Int

The response's http status code.

newDescribeScheduledInstanceAvailabilityResponse

Arguments

:: Int
$sel:httpStatus:DescribeScheduledInstanceAvailabilityResponse'

-> DescribeScheduledInstanceAvailabilityResponse

Create a value of DescribeScheduledInstanceAvailabilityResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:scheduledInstanceAvailabilitySet:DescribeScheduledInstanceAvailabilityResponse', describeScheduledInstanceAvailabilityResponse_scheduledInstanceAvailabilitySet - Information about the available Scheduled Instances.

$sel:nextToken:DescribeScheduledInstanceAvailability', describeScheduledInstanceAvailabilityResponse_nextToken - The token required to retrieve the next set of results. This value is null when there are no more results to return.

$sel:httpStatus:DescribeScheduledInstanceAvailabilityResponse', describeScheduledInstanceAvailabilityResponse_httpStatus - The response's http status code.

Response Lenses

describeScheduledInstanceAvailabilityResponse_scheduledInstanceAvailabilitySet :: Lens' DescribeScheduledInstanceAvailabilityResponse (Maybe [ScheduledInstanceAvailability])

Information about the available Scheduled Instances.

describeScheduledInstanceAvailabilityResponse_nextToken :: Lens' DescribeScheduledInstanceAvailabilityResponse (Maybe Text)

The token required to retrieve the next set of results. This value is null when there are no more results to return.

describeScheduledInstanceAvailabilityResponse_httpStatus :: Lens' DescribeScheduledInstanceAvailabilityResponse Int

The response's http status code.