Network.AWS.EC2.DescribeSecurityGroupReferences

Description

VPC only

Describes the VPCs on the other side of a VPC peering connection that are referencing the security groups you've specified in this request.

Creating a Request

data DescribeSecurityGroupReferences

See: newDescribeSecurityGroupReferences smart constructor.

Constructors

DescribeSecurityGroupReferences'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

groupId :: [Text]

The IDs of the security groups in your account.

newDescribeSecurityGroupReferences :: DescribeSecurityGroupReferences

Create a value of DescribeSecurityGroupReferences with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:DescribeSecurityGroupReferences', describeSecurityGroupReferences_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:groupId:DescribeSecurityGroupReferences', describeSecurityGroupReferences_groupId - The IDs of the security groups in your account.

Request Lenses

describeSecurityGroupReferences_dryRun :: Lens' DescribeSecurityGroupReferences (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeSecurityGroupReferences_groupId :: Lens' DescribeSecurityGroupReferences [Text]

The IDs of the security groups in your account.

Destructuring the Response

data DescribeSecurityGroupReferencesResponse

See: newDescribeSecurityGroupReferencesResponse smart constructor.

Constructors

DescribeSecurityGroupReferencesResponse'

Fields

securityGroupReferenceSet :: Maybe [SecurityGroupReference]

Information about the VPCs with the referencing security groups.

httpStatus :: Int

The response's http status code.

newDescribeSecurityGroupReferencesResponse

Arguments

:: Int
$sel:httpStatus:DescribeSecurityGroupReferencesResponse'

-> DescribeSecurityGroupReferencesResponse

Create a value of DescribeSecurityGroupReferencesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:securityGroupReferenceSet:DescribeSecurityGroupReferencesResponse', describeSecurityGroupReferencesResponse_securityGroupReferenceSet - Information about the VPCs with the referencing security groups.

$sel:httpStatus:DescribeSecurityGroupReferencesResponse', describeSecurityGroupReferencesResponse_httpStatus - The response's http status code.

Response Lenses

describeSecurityGroupReferencesResponse_securityGroupReferenceSet :: Lens' DescribeSecurityGroupReferencesResponse (Maybe [SecurityGroupReference])

Information about the VPCs with the referencing security groups.

describeSecurityGroupReferencesResponse_httpStatus :: Lens' DescribeSecurityGroupReferencesResponse Int

The response's http status code.