Network.AWS.EC2.DescribeSecurityGroups

Description

Describes the specified security groups or all of your security groups.

A security group is for use with instances either in the EC2-Classic platform or in a specific VPC. For more information, see Amazon EC2 security groups in the Amazon Elastic Compute Cloud User Guide and Security groups for your VPC in the Amazon Virtual Private Cloud User Guide.

This operation returns paginated results.

Creating a Request

data DescribeSecurityGroups

See: newDescribeSecurityGroups smart constructor.

Constructors

DescribeSecurityGroups'

Fields

filters :: Maybe [Filter]

The filters. If using multiple filters for rules, the results include security groups for which any combination of rules - not necessarily a single rule - match all filters.

description - The description of the security group.

egress.ip-permission.cidr - An IPv4 CIDR block for an outbound security group rule.

egress.ip-permission.from-port - For an outbound rule, the start of port range for the TCP and UDP protocols, or an ICMP type number.

egress.ip-permission.group-id - The ID of a security group that has been referenced in an outbound security group rule.

egress.ip-permission.group-name - The name of a security group that is referenced in an outbound security group rule.

egress.ip-permission.ipv6-cidr - An IPv6 CIDR block for an outbound security group rule.

egress.ip-permission.prefix-list-id - The ID of a prefix list to which a security group rule allows outbound access.

egress.ip-permission.protocol - The IP protocol for an outbound security group rule (tcp | udp | icmp, a protocol number, or

1 for all protocols).

egress.ip-permission.to-port - For an outbound rule, the end of port range for the TCP and UDP protocols, or an ICMP code.

egress.ip-permission.user-id - The ID of an Amazon Web Services account that has been referenced in an outbound security group rule.

group-id - The ID of the security group.

group-name - The name of the security group.

ip-permission.cidr - An IPv4 CIDR block for an inbound security group rule.

ip-permission.from-port - For an inbound rule, the start of port range for the TCP and UDP protocols, or an ICMP type number.

ip-permission.group-id - The ID of a security group that has been referenced in an inbound security group rule.

ip-permission.group-name - The name of a security group that is referenced in an inbound security group rule.

ip-permission.ipv6-cidr - An IPv6 CIDR block for an inbound security group rule.

ip-permission.prefix-list-id - The ID of a prefix list from which a security group rule allows inbound access.

ip-permission.protocol - The IP protocol for an inbound security group rule (tcp | udp | icmp, a protocol number, or -1 for all protocols).

ip-permission.to-port - For an inbound rule, the end of port range for the TCP and UDP protocols, or an ICMP code.

ip-permission.user-id - The ID of an Amazon Web Services account that has been referenced in an inbound security group rule.

owner-id - The Amazon Web Services account ID of the owner of the security group.

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

vpc-id - The ID of the VPC specified when the security group was created.

groupNames :: Maybe [Text]

EC2-Classic and default VPC only

The names of the security groups. You can specify either the security group name or the security group ID. For security groups in a nondefault VPC, use the group-name filter to describe security groups by name.

Default: Describes all of your security groups.

groupIds :: Maybe [Text]

The IDs of the security groups. Required for security groups in a nondefault VPC.

Default: Describes all of your security groups.

nextToken :: Maybe Text

The token to request the next page of results.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

maxResults :: Maybe Natural

The maximum number of results to return in a single call. To retrieve the remaining results, make another request with the returned NextToken value. This value can be between 5 and 1000. If this parameter is not specified, then all results are returned.

newDescribeSecurityGroups :: DescribeSecurityGroups

Create a value of DescribeSecurityGroups with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:DescribeSecurityGroups', describeSecurityGroups_filters - The filters. If using multiple filters for rules, the results include security groups for which any combination of rules - not necessarily a single rule - match all filters.

description - The description of the security group.

egress.ip-permission.cidr - An IPv4 CIDR block for an outbound security group rule.

egress.ip-permission.from-port - For an outbound rule, the start of port range for the TCP and UDP protocols, or an ICMP type number.

egress.ip-permission.group-id - The ID of a security group that has been referenced in an outbound security group rule.

egress.ip-permission.group-name - The name of a security group that is referenced in an outbound security group rule.

egress.ip-permission.ipv6-cidr - An IPv6 CIDR block for an outbound security group rule.

egress.ip-permission.prefix-list-id - The ID of a prefix list to which a security group rule allows outbound access.

egress.ip-permission.protocol - The IP protocol for an outbound security group rule (tcp | udp | icmp, a protocol number, or

1 for all protocols).

egress.ip-permission.to-port - For an outbound rule, the end of port range for the TCP and UDP protocols, or an ICMP code.

egress.ip-permission.user-id - The ID of an Amazon Web Services account that has been referenced in an outbound security group rule.

group-id - The ID of the security group.

group-name - The name of the security group.

ip-permission.cidr - An IPv4 CIDR block for an inbound security group rule.

ip-permission.from-port - For an inbound rule, the start of port range for the TCP and UDP protocols, or an ICMP type number.

ip-permission.group-id - The ID of a security group that has been referenced in an inbound security group rule.

ip-permission.group-name - The name of a security group that is referenced in an inbound security group rule.

ip-permission.ipv6-cidr - An IPv6 CIDR block for an inbound security group rule.

ip-permission.prefix-list-id - The ID of a prefix list from which a security group rule allows inbound access.

ip-permission.protocol - The IP protocol for an inbound security group rule (tcp | udp | icmp, a protocol number, or -1 for all protocols).

ip-permission.to-port - For an inbound rule, the end of port range for the TCP and UDP protocols, or an ICMP code.

ip-permission.user-id - The ID of an Amazon Web Services account that has been referenced in an inbound security group rule.

owner-id - The Amazon Web Services account ID of the owner of the security group.

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

vpc-id - The ID of the VPC specified when the security group was created.

$sel:groupNames:DescribeSecurityGroups', describeSecurityGroups_groupNames - [EC2-Classic and default VPC only] The names of the security groups. You can specify either the security group name or the security group ID. For security groups in a nondefault VPC, use the group-name filter to describe security groups by name.

Default: Describes all of your security groups.

$sel:groupIds:DescribeSecurityGroups', describeSecurityGroups_groupIds - The IDs of the security groups. Required for security groups in a nondefault VPC.

Default: Describes all of your security groups.

$sel:nextToken:DescribeSecurityGroups', describeSecurityGroups_nextToken - The token to request the next page of results.

$sel:dryRun:DescribeSecurityGroups', describeSecurityGroups_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:maxResults:DescribeSecurityGroups', describeSecurityGroups_maxResults - The maximum number of results to return in a single call. To retrieve the remaining results, make another request with the returned NextToken value. This value can be between 5 and 1000. If this parameter is not specified, then all results are returned.

Request Lenses

describeSecurityGroups_filters :: Lens' DescribeSecurityGroups (Maybe [Filter])

The filters. If using multiple filters for rules, the results include security groups for which any combination of rules - not necessarily a single rule - match all filters.

description - The description of the security group.

egress.ip-permission.cidr - An IPv4 CIDR block for an outbound security group rule.

egress.ip-permission.from-port - For an outbound rule, the start of port range for the TCP and UDP protocols, or an ICMP type number.

egress.ip-permission.group-id - The ID of a security group that has been referenced in an outbound security group rule.

egress.ip-permission.group-name - The name of a security group that is referenced in an outbound security group rule.

egress.ip-permission.ipv6-cidr - An IPv6 CIDR block for an outbound security group rule.

egress.ip-permission.prefix-list-id - The ID of a prefix list to which a security group rule allows outbound access.

egress.ip-permission.protocol - The IP protocol for an outbound security group rule (tcp | udp | icmp, a protocol number, or

1 for all protocols).

egress.ip-permission.to-port - For an outbound rule, the end of port range for the TCP and UDP protocols, or an ICMP code.

egress.ip-permission.user-id - The ID of an Amazon Web Services account that has been referenced in an outbound security group rule.

group-id - The ID of the security group.

group-name - The name of the security group.

ip-permission.cidr - An IPv4 CIDR block for an inbound security group rule.

ip-permission.from-port - For an inbound rule, the start of port range for the TCP and UDP protocols, or an ICMP type number.

ip-permission.group-id - The ID of a security group that has been referenced in an inbound security group rule.

ip-permission.group-name - The name of a security group that is referenced in an inbound security group rule.

ip-permission.ipv6-cidr - An IPv6 CIDR block for an inbound security group rule.

ip-permission.prefix-list-id - The ID of a prefix list from which a security group rule allows inbound access.

ip-permission.protocol - The IP protocol for an inbound security group rule (tcp | udp | icmp, a protocol number, or -1 for all protocols).

ip-permission.to-port - For an inbound rule, the end of port range for the TCP and UDP protocols, or an ICMP code.

ip-permission.user-id - The ID of an Amazon Web Services account that has been referenced in an inbound security group rule.

owner-id - The Amazon Web Services account ID of the owner of the security group.

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

vpc-id - The ID of the VPC specified when the security group was created.

describeSecurityGroups_groupNames :: Lens' DescribeSecurityGroups (Maybe [Text])

EC2-Classic and default VPC only

The names of the security groups. You can specify either the security group name or the security group ID. For security groups in a nondefault VPC, use the group-name filter to describe security groups by name.

Default: Describes all of your security groups.

describeSecurityGroups_groupIds :: Lens' DescribeSecurityGroups (Maybe [Text])

The IDs of the security groups. Required for security groups in a nondefault VPC.

Default: Describes all of your security groups.

describeSecurityGroups_nextToken :: Lens' DescribeSecurityGroups (Maybe Text)

The token to request the next page of results.

describeSecurityGroups_dryRun :: Lens' DescribeSecurityGroups (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeSecurityGroups_maxResults :: Lens' DescribeSecurityGroups (Maybe Natural)

The maximum number of results to return in a single call. To retrieve the remaining results, make another request with the returned NextToken value. This value can be between 5 and 1000. If this parameter is not specified, then all results are returned.

Destructuring the Response

data DescribeSecurityGroupsResponse

See: newDescribeSecurityGroupsResponse smart constructor.

Constructors

DescribeSecurityGroupsResponse'

Fields

securityGroups :: Maybe [SecurityGroup]

Information about the security groups.

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

httpStatus :: Int

The response's http status code.

newDescribeSecurityGroupsResponse

Arguments

:: Int
$sel:httpStatus:DescribeSecurityGroupsResponse'

-> DescribeSecurityGroupsResponse

Create a value of DescribeSecurityGroupsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:securityGroups:DescribeSecurityGroupsResponse', describeSecurityGroupsResponse_securityGroups - Information about the security groups.

$sel:nextToken:DescribeSecurityGroups', describeSecurityGroupsResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:httpStatus:DescribeSecurityGroupsResponse', describeSecurityGroupsResponse_httpStatus - The response's http status code.

Response Lenses

describeSecurityGroupsResponse_securityGroups :: Lens' DescribeSecurityGroupsResponse (Maybe [SecurityGroup])

Information about the security groups.

describeSecurityGroupsResponse_nextToken :: Lens' DescribeSecurityGroupsResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

describeSecurityGroupsResponse_httpStatus :: Lens' DescribeSecurityGroupsResponse Int

The response's http status code.