Network.AWS.EC2.DescribeSnapshotAttribute

Description

Describes the specified attribute of the specified snapshot. You can specify only one attribute at a time.

For more information about EBS snapshots, see Amazon EBS snapshots in the Amazon Elastic Compute Cloud User Guide.

Creating a Request

data DescribeSnapshotAttribute

See: newDescribeSnapshotAttribute smart constructor.

Constructors

DescribeSnapshotAttribute'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

attribute :: SnapshotAttributeName

The snapshot attribute you would like to view.

snapshotId :: Text

The ID of the EBS snapshot.

newDescribeSnapshotAttribute

Arguments

:: SnapshotAttributeName
$sel:attribute:DescribeSnapshotAttribute'

-> Text
$sel:snapshotId:DescribeSnapshotAttribute'

-> DescribeSnapshotAttribute

Create a value of DescribeSnapshotAttribute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:DescribeSnapshotAttribute', describeSnapshotAttribute_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:attribute:DescribeSnapshotAttribute', describeSnapshotAttribute_attribute - The snapshot attribute you would like to view.

$sel:snapshotId:DescribeSnapshotAttribute', describeSnapshotAttribute_snapshotId - The ID of the EBS snapshot.

Request Lenses

describeSnapshotAttribute_dryRun :: Lens' DescribeSnapshotAttribute (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeSnapshotAttribute_attribute :: Lens' DescribeSnapshotAttribute SnapshotAttributeName

The snapshot attribute you would like to view.

describeSnapshotAttribute_snapshotId :: Lens' DescribeSnapshotAttribute Text

The ID of the EBS snapshot.

Destructuring the Response

data DescribeSnapshotAttributeResponse

See: newDescribeSnapshotAttributeResponse smart constructor.

Constructors

DescribeSnapshotAttributeResponse'

Fields

productCodes :: Maybe [ProductCode]

The product codes.

createVolumePermissions :: Maybe [CreateVolumePermission]

The users and groups that have the permissions for creating volumes from the snapshot.

snapshotId :: Maybe Text

The ID of the EBS snapshot.

httpStatus :: Int

The response's http status code.

newDescribeSnapshotAttributeResponse

Arguments

:: Int
$sel:httpStatus:DescribeSnapshotAttributeResponse'

-> DescribeSnapshotAttributeResponse

Create a value of DescribeSnapshotAttributeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:productCodes:DescribeSnapshotAttributeResponse', describeSnapshotAttributeResponse_productCodes - The product codes.

$sel:createVolumePermissions:DescribeSnapshotAttributeResponse', describeSnapshotAttributeResponse_createVolumePermissions - The users and groups that have the permissions for creating volumes from the snapshot.

$sel:snapshotId:DescribeSnapshotAttribute', describeSnapshotAttributeResponse_snapshotId - The ID of the EBS snapshot.

$sel:httpStatus:DescribeSnapshotAttributeResponse', describeSnapshotAttributeResponse_httpStatus - The response's http status code.

Response Lenses

describeSnapshotAttributeResponse_productCodes :: Lens' DescribeSnapshotAttributeResponse (Maybe [ProductCode])

The product codes.

describeSnapshotAttributeResponse_createVolumePermissions :: Lens' DescribeSnapshotAttributeResponse (Maybe [CreateVolumePermission])

The users and groups that have the permissions for creating volumes from the snapshot.

describeSnapshotAttributeResponse_snapshotId :: Lens' DescribeSnapshotAttributeResponse (Maybe Text)

The ID of the EBS snapshot.

describeSnapshotAttributeResponse_httpStatus :: Lens' DescribeSnapshotAttributeResponse Int

The response's http status code.