Network.AWS.EC2.DescribeSnapshots

Description

Describes the specified EBS snapshots available to you or all of the EBS snapshots available to you.

The snapshots available to you include public snapshots, private snapshots that you own, and private snapshots owned by other Amazon Web Services accounts for which you have explicit create volume permissions.

The create volume permissions fall into the following categories:

public: The owner of the snapshot granted create volume permissions for the snapshot to the all group. All Amazon Web Services accounts have create volume permissions for these snapshots.

explicit: The owner of the snapshot granted create volume permissions to a specific Amazon Web Services account.

implicit: An Amazon Web Services account has implicit create volume permissions for all snapshots it owns.

The list of snapshots returned can be filtered by specifying snapshot IDs, snapshot owners, or Amazon Web Services accounts with create volume permissions. If no options are specified, Amazon EC2 returns all snapshots for which you have create volume permissions.

If you specify one or more snapshot IDs, only snapshots that have the specified IDs are returned. If you specify an invalid snapshot ID, an error is returned. If you specify a snapshot ID for which you do not have access, it is not included in the returned results.

If you specify one or more snapshot owners using the OwnerIds option, only snapshots from the specified owners and for which you have access are returned. The results can include the Amazon Web Services account IDs of the specified owners, amazon for snapshots owned by Amazon, or self for snapshots that you own.

If you specify a list of restorable users, only snapshots with create snapshot permissions for those users are returned. You can specify Amazon Web Services account IDs (if you own the snapshots), self for snapshots for which you own or have explicit permissions, or all for public snapshots.

If you are describing a long list of snapshots, we recommend that you paginate the output to make the list more manageable. The MaxResults parameter sets the maximum number of results returned in a single page. If the list of results exceeds your MaxResults value, then that number of results is returned along with a NextToken value that can be passed to a subsequent DescribeSnapshots request to retrieve the remaining results.

To get the state of fast snapshot restores for a snapshot, use DescribeFastSnapshotRestores.

For more information about EBS snapshots, see Amazon EBS snapshots in the Amazon Elastic Compute Cloud User Guide.

This operation returns paginated results.

Creating a Request

data DescribeSnapshots

See: newDescribeSnapshots smart constructor.

Constructors

DescribeSnapshots'

Fields

ownerIds :: Maybe [Text]

Scopes the results to snapshots with the specified owners. You can specify a combination of Amazon Web Services account IDs, self, and amazon.

nextToken :: Maybe Text

The NextToken value returned from a previous paginated DescribeSnapshots request where MaxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the NextToken value. This value is null when there are no more results to return.

maxResults :: Maybe Int

The maximum number of snapshot results returned by DescribeSnapshots in paginated output. When this parameter is used, DescribeSnapshots only returns MaxResults results in a single page along with a NextToken response element. The remaining results of the initial request can be seen by sending another DescribeSnapshots request with the returned NextToken value. This value can be between 5 and 1,000; if MaxResults is given a value larger than 1,000, only 1,000 results are returned. If this parameter is not used, then DescribeSnapshots returns all results. You cannot specify this parameter and the snapshot IDs parameter in the same request.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

restorableByUserIds :: Maybe [Text]

The IDs of the Amazon Web Services accounts that can create volumes from the snapshot.

snapshotIds :: Maybe [Text]

The snapshot IDs.

Default: Describes the snapshots for which you have create volume permissions.

filters :: Maybe [Filter]

The filters.

description - A description of the snapshot.

encrypted - Indicates whether the snapshot is encrypted (true | false)

owner-alias - The owner alias, from an Amazon-maintained list (amazon). This is not the user-configured Amazon Web Services account alias set using the IAM console. We recommend that you use the related parameter instead of this filter.

owner-id - The Amazon Web Services account ID of the owner. We recommend that you use the related parameter instead of this filter.

progress - The progress of the snapshot, as a percentage (for example, 80%).

snapshot-id - The snapshot ID.

start-time - The time stamp when the snapshot was initiated.

status - The status of the snapshot (pending | completed | error).

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

volume-id - The ID of the volume the snapshot is for.

volume-size - The size of the volume, in GiB.

newDescribeSnapshots :: DescribeSnapshots

Create a value of DescribeSnapshots with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ownerIds:DescribeSnapshots', describeSnapshots_ownerIds - Scopes the results to snapshots with the specified owners. You can specify a combination of Amazon Web Services account IDs, self, and amazon.

$sel:nextToken:DescribeSnapshots', describeSnapshots_nextToken - The NextToken value returned from a previous paginated DescribeSnapshots request where MaxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the NextToken value. This value is null when there are no more results to return.

$sel:maxResults:DescribeSnapshots', describeSnapshots_maxResults - The maximum number of snapshot results returned by DescribeSnapshots in paginated output. When this parameter is used, DescribeSnapshots only returns MaxResults results in a single page along with a NextToken response element. The remaining results of the initial request can be seen by sending another DescribeSnapshots request with the returned NextToken value. This value can be between 5 and 1,000; if MaxResults is given a value larger than 1,000, only 1,000 results are returned. If this parameter is not used, then DescribeSnapshots returns all results. You cannot specify this parameter and the snapshot IDs parameter in the same request.

$sel:dryRun:DescribeSnapshots', describeSnapshots_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:restorableByUserIds:DescribeSnapshots', describeSnapshots_restorableByUserIds - The IDs of the Amazon Web Services accounts that can create volumes from the snapshot.

$sel:snapshotIds:DescribeSnapshots', describeSnapshots_snapshotIds - The snapshot IDs.

Default: Describes the snapshots for which you have create volume permissions.

$sel:filters:DescribeSnapshots', describeSnapshots_filters - The filters.

description - A description of the snapshot.

encrypted - Indicates whether the snapshot is encrypted (true | false)

owner-alias - The owner alias, from an Amazon-maintained list (amazon). This is not the user-configured Amazon Web Services account alias set using the IAM console. We recommend that you use the related parameter instead of this filter.

owner-id - The Amazon Web Services account ID of the owner. We recommend that you use the related parameter instead of this filter.

progress - The progress of the snapshot, as a percentage (for example, 80%).

snapshot-id - The snapshot ID.

start-time - The time stamp when the snapshot was initiated.

status - The status of the snapshot (pending | completed | error).

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

volume-id - The ID of the volume the snapshot is for.

volume-size - The size of the volume, in GiB.

Request Lenses

describeSnapshots_ownerIds :: Lens' DescribeSnapshots (Maybe [Text])

Scopes the results to snapshots with the specified owners. You can specify a combination of Amazon Web Services account IDs, self, and amazon.

describeSnapshots_nextToken :: Lens' DescribeSnapshots (Maybe Text)

The NextToken value returned from a previous paginated DescribeSnapshots request where MaxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the NextToken value. This value is null when there are no more results to return.

describeSnapshots_maxResults :: Lens' DescribeSnapshots (Maybe Int)

The maximum number of snapshot results returned by DescribeSnapshots in paginated output. When this parameter is used, DescribeSnapshots only returns MaxResults results in a single page along with a NextToken response element. The remaining results of the initial request can be seen by sending another DescribeSnapshots request with the returned NextToken value. This value can be between 5 and 1,000; if MaxResults is given a value larger than 1,000, only 1,000 results are returned. If this parameter is not used, then DescribeSnapshots returns all results. You cannot specify this parameter and the snapshot IDs parameter in the same request.

describeSnapshots_dryRun :: Lens' DescribeSnapshots (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeSnapshots_restorableByUserIds :: Lens' DescribeSnapshots (Maybe [Text])

The IDs of the Amazon Web Services accounts that can create volumes from the snapshot.

describeSnapshots_snapshotIds :: Lens' DescribeSnapshots (Maybe [Text])

The snapshot IDs.

Default: Describes the snapshots for which you have create volume permissions.

describeSnapshots_filters :: Lens' DescribeSnapshots (Maybe [Filter])

The filters.

description - A description of the snapshot.

encrypted - Indicates whether the snapshot is encrypted (true | false)

owner-alias - The owner alias, from an Amazon-maintained list (amazon). This is not the user-configured Amazon Web Services account alias set using the IAM console. We recommend that you use the related parameter instead of this filter.

owner-id - The Amazon Web Services account ID of the owner. We recommend that you use the related parameter instead of this filter.

progress - The progress of the snapshot, as a percentage (for example, 80%).

snapshot-id - The snapshot ID.

start-time - The time stamp when the snapshot was initiated.

status - The status of the snapshot (pending | completed | error).

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

volume-id - The ID of the volume the snapshot is for.

volume-size - The size of the volume, in GiB.

Destructuring the Response

data DescribeSnapshotsResponse

See: newDescribeSnapshotsResponse smart constructor.

Constructors

DescribeSnapshotsResponse'

Fields

snapshots :: Maybe [Snapshot]

Information about the snapshots.

nextToken :: Maybe Text

The NextToken value to include in a future DescribeSnapshots request. When the results of a DescribeSnapshots request exceed MaxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

httpStatus :: Int

The response's http status code.

newDescribeSnapshotsResponse

Arguments

:: Int
$sel:httpStatus:DescribeSnapshotsResponse'

-> DescribeSnapshotsResponse

Create a value of DescribeSnapshotsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:snapshots:DescribeSnapshotsResponse', describeSnapshotsResponse_snapshots - Information about the snapshots.

$sel:nextToken:DescribeSnapshots', describeSnapshotsResponse_nextToken - The NextToken value to include in a future DescribeSnapshots request. When the results of a DescribeSnapshots request exceed MaxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

$sel:httpStatus:DescribeSnapshotsResponse', describeSnapshotsResponse_httpStatus - The response's http status code.

Response Lenses

describeSnapshotsResponse_snapshots :: Lens' DescribeSnapshotsResponse (Maybe [Snapshot])

Information about the snapshots.

describeSnapshotsResponse_nextToken :: Lens' DescribeSnapshotsResponse (Maybe Text)

The NextToken value to include in a future DescribeSnapshots request. When the results of a DescribeSnapshots request exceed MaxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

describeSnapshotsResponse_httpStatus :: Lens' DescribeSnapshotsResponse Int

The response's http status code.