Network.AWS.EC2.DescribeSpotFleetRequests

Description

Describes your Spot Fleet requests.

Spot Fleet requests are deleted 48 hours after they are canceled and their instances are terminated.

This operation returns paginated results.

Creating a Request

data DescribeSpotFleetRequests

Contains the parameters for DescribeSpotFleetRequests.

See: newDescribeSpotFleetRequests smart constructor.

Constructors

DescribeSpotFleetRequests'

Fields

nextToken :: Maybe Text

The token for the next set of results.

maxResults :: Maybe Int

The maximum number of results to return in a single call. Specify a value between 1 and 1000. The default value is 1000. To retrieve the remaining results, make another call with the returned NextToken value.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

spotFleetRequestIds :: Maybe [Text]

The IDs of the Spot Fleet requests.

newDescribeSpotFleetRequests :: DescribeSpotFleetRequests

Create a value of DescribeSpotFleetRequests with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeSpotFleetRequests', describeSpotFleetRequests_nextToken - The token for the next set of results.

$sel:maxResults:DescribeSpotFleetRequests', describeSpotFleetRequests_maxResults - The maximum number of results to return in a single call. Specify a value between 1 and 1000. The default value is 1000. To retrieve the remaining results, make another call with the returned NextToken value.

$sel:dryRun:DescribeSpotFleetRequests', describeSpotFleetRequests_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:spotFleetRequestIds:DescribeSpotFleetRequests', describeSpotFleetRequests_spotFleetRequestIds - The IDs of the Spot Fleet requests.

Request Lenses

describeSpotFleetRequests_nextToken :: Lens' DescribeSpotFleetRequests (Maybe Text)

The token for the next set of results.

describeSpotFleetRequests_maxResults :: Lens' DescribeSpotFleetRequests (Maybe Int)

The maximum number of results to return in a single call. Specify a value between 1 and 1000. The default value is 1000. To retrieve the remaining results, make another call with the returned NextToken value.

describeSpotFleetRequests_dryRun :: Lens' DescribeSpotFleetRequests (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeSpotFleetRequests_spotFleetRequestIds :: Lens' DescribeSpotFleetRequests (Maybe [Text])

The IDs of the Spot Fleet requests.

Destructuring the Response

data DescribeSpotFleetRequestsResponse

Contains the output of DescribeSpotFleetRequests.

See: newDescribeSpotFleetRequestsResponse smart constructor.

Constructors

DescribeSpotFleetRequestsResponse'

Fields

nextToken :: Maybe Text

The token required to retrieve the next set of results. This value is null when there are no more results to return.

spotFleetRequestConfigs :: Maybe [SpotFleetRequestConfig]

Information about the configuration of your Spot Fleet.

httpStatus :: Int

The response's http status code.

newDescribeSpotFleetRequestsResponse

Arguments

:: Int
$sel:httpStatus:DescribeSpotFleetRequestsResponse'

-> DescribeSpotFleetRequestsResponse

Create a value of DescribeSpotFleetRequestsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeSpotFleetRequests', describeSpotFleetRequestsResponse_nextToken - The token required to retrieve the next set of results. This value is null when there are no more results to return.

$sel:spotFleetRequestConfigs:DescribeSpotFleetRequestsResponse', describeSpotFleetRequestsResponse_spotFleetRequestConfigs - Information about the configuration of your Spot Fleet.

$sel:httpStatus:DescribeSpotFleetRequestsResponse', describeSpotFleetRequestsResponse_httpStatus - The response's http status code.

Response Lenses

describeSpotFleetRequestsResponse_nextToken :: Lens' DescribeSpotFleetRequestsResponse (Maybe Text)

The token required to retrieve the next set of results. This value is null when there are no more results to return.

describeSpotFleetRequestsResponse_spotFleetRequestConfigs :: Lens' DescribeSpotFleetRequestsResponse (Maybe [SpotFleetRequestConfig])

Information about the configuration of your Spot Fleet.

describeSpotFleetRequestsResponse_httpStatus :: Lens' DescribeSpotFleetRequestsResponse Int

The response's http status code.