Network.AWS.EC2.DescribeSpotInstanceRequests

Description

Describes the specified Spot Instance requests.

You can use DescribeSpotInstanceRequests to find a running Spot Instance by examining the response. If the status of the Spot Instance is fulfilled, the instance ID appears in the response and contains the identifier of the instance. Alternatively, you can use DescribeInstances with a filter to look for instances where the instance lifecycle is spot.

We recommend that you set MaxResults to a value between 5 and 1000 to limit the number of results returned. This paginates the output, which makes the list more manageable and returns the results faster. If the list of results exceeds your MaxResults value, then that number of results is returned along with a NextToken value that can be passed to a subsequent DescribeSpotInstanceRequests request to retrieve the remaining results.

Spot Instance requests are deleted four hours after they are canceled and their instances are terminated.

This operation returns paginated results.

Creating a Request

data DescribeSpotInstanceRequests

Contains the parameters for DescribeSpotInstanceRequests.

See: newDescribeSpotInstanceRequests smart constructor.

Constructors

DescribeSpotInstanceRequests'

Fields

nextToken :: Maybe Text

The token to request the next set of results. This value is null when there are no more results to return.

maxResults :: Maybe Int

The maximum number of results to return in a single call. Specify a value between 5 and 1000. To retrieve the remaining results, make another call with the returned NextToken value.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

spotInstanceRequestIds :: Maybe [Text]

One or more Spot Instance request IDs.

filters :: Maybe [Filter]

One or more filters.

availability-zone-group - The Availability Zone group.

create-time - The time stamp when the Spot Instance request was created.

fault-code - The fault code related to the request.

fault-message - The fault message related to the request.

instance-id - The ID of the instance that fulfilled the request.

launch-group - The Spot Instance launch group.

launch.block-device-mapping.delete-on-termination - Indicates whether the EBS volume is deleted on instance termination.

launch.block-device-mapping.device-name - The device name for the volume in the block device mapping (for example, /dev/sdh or xvdh).

launch.block-device-mapping.snapshot-id - The ID of the snapshot for the EBS volume.

launch.block-device-mapping.volume-size - The size of the EBS volume, in GiB.

launch.block-device-mapping.volume-type - The type of EBS volume: gp2 for General Purpose SSD, io1 or io2 for Provisioned IOPS SSD, st1 for Throughput Optimized HDD, sc1for Cold HDD, or standard for Magnetic.

launch.group-id - The ID of the security group for the instance.

launch.group-name - The name of the security group for the instance.

launch.image-id - The ID of the AMI.

launch.instance-type - The type of instance (for example, m3.medium).

launch.kernel-id - The kernel ID.

launch.key-name - The name of the key pair the instance launched with.

launch.monitoring-enabled - Whether detailed monitoring is enabled for the Spot Instance.

launch.ramdisk-id - The RAM disk ID.

launched-availability-zone - The Availability Zone in which the request is launched.

network-interface.addresses.primary - Indicates whether the IP address is the primary private IP address.

network-interface.delete-on-termination - Indicates whether the network interface is deleted when the instance is terminated.

network-interface.description - A description of the network interface.

network-interface.device-index - The index of the device for the network interface attachment on the instance.

network-interface.group-id - The ID of the security group associated with the network interface.

network-interface.network-interface-id - The ID of the network interface.

network-interface.private-ip-address - The primary private IP address of the network interface.

network-interface.subnet-id - The ID of the subnet for the instance.

product-description - The product description associated with the instance (Linux/UNIX | Windows).

spot-instance-request-id - The Spot Instance request ID.

spot-price - The maximum hourly price for any Spot Instance launched to fulfill the request.

state - The state of the Spot Instance request (open | active | closed | cancelled | failed). Spot request status information can help you track your Amazon EC2 Spot Instance requests. For more information, see Spot request status in the Amazon EC2 User Guide for Linux Instances.

status-code - The short code describing the most recent evaluation of your Spot Instance request.

status-message - The message explaining the status of the Spot Instance request.

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

type - The type of Spot Instance request (one-time | persistent).

valid-from - The start date of the request.

valid-until - The end date of the request.

newDescribeSpotInstanceRequests :: DescribeSpotInstanceRequests

Create a value of DescribeSpotInstanceRequests with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeSpotInstanceRequests', describeSpotInstanceRequests_nextToken - The token to request the next set of results. This value is null when there are no more results to return.

$sel:maxResults:DescribeSpotInstanceRequests', describeSpotInstanceRequests_maxResults - The maximum number of results to return in a single call. Specify a value between 5 and 1000. To retrieve the remaining results, make another call with the returned NextToken value.

$sel:dryRun:DescribeSpotInstanceRequests', describeSpotInstanceRequests_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:spotInstanceRequestIds:DescribeSpotInstanceRequests', describeSpotInstanceRequests_spotInstanceRequestIds - One or more Spot Instance request IDs.

$sel:filters:DescribeSpotInstanceRequests', describeSpotInstanceRequests_filters - One or more filters.

availability-zone-group - The Availability Zone group.

create-time - The time stamp when the Spot Instance request was created.

fault-code - The fault code related to the request.

fault-message - The fault message related to the request.

instance-id - The ID of the instance that fulfilled the request.

launch-group - The Spot Instance launch group.

launch.block-device-mapping.delete-on-termination - Indicates whether the EBS volume is deleted on instance termination.

launch.block-device-mapping.device-name - The device name for the volume in the block device mapping (for example, /dev/sdh or xvdh).

launch.block-device-mapping.snapshot-id - The ID of the snapshot for the EBS volume.

launch.block-device-mapping.volume-size - The size of the EBS volume, in GiB.

launch.block-device-mapping.volume-type - The type of EBS volume: gp2 for General Purpose SSD, io1 or io2 for Provisioned IOPS SSD, st1 for Throughput Optimized HDD, sc1for Cold HDD, or standard for Magnetic.

launch.group-id - The ID of the security group for the instance.

launch.group-name - The name of the security group for the instance.

launch.image-id - The ID of the AMI.

launch.instance-type - The type of instance (for example, m3.medium).

launch.kernel-id - The kernel ID.

launch.key-name - The name of the key pair the instance launched with.

launch.monitoring-enabled - Whether detailed monitoring is enabled for the Spot Instance.

launch.ramdisk-id - The RAM disk ID.

launched-availability-zone - The Availability Zone in which the request is launched.

network-interface.addresses.primary - Indicates whether the IP address is the primary private IP address.

network-interface.delete-on-termination - Indicates whether the network interface is deleted when the instance is terminated.

network-interface.description - A description of the network interface.

network-interface.device-index - The index of the device for the network interface attachment on the instance.

network-interface.group-id - The ID of the security group associated with the network interface.

network-interface.network-interface-id - The ID of the network interface.

network-interface.private-ip-address - The primary private IP address of the network interface.

network-interface.subnet-id - The ID of the subnet for the instance.

product-description - The product description associated with the instance (Linux/UNIX | Windows).

spot-instance-request-id - The Spot Instance request ID.

spot-price - The maximum hourly price for any Spot Instance launched to fulfill the request.

state - The state of the Spot Instance request (open | active | closed | cancelled | failed). Spot request status information can help you track your Amazon EC2 Spot Instance requests. For more information, see Spot request status in the Amazon EC2 User Guide for Linux Instances.

status-code - The short code describing the most recent evaluation of your Spot Instance request.

status-message - The message explaining the status of the Spot Instance request.

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

type - The type of Spot Instance request (one-time | persistent).

valid-from - The start date of the request.

valid-until - The end date of the request.

Request Lenses

describeSpotInstanceRequests_nextToken :: Lens' DescribeSpotInstanceRequests (Maybe Text)

The token to request the next set of results. This value is null when there are no more results to return.

describeSpotInstanceRequests_maxResults :: Lens' DescribeSpotInstanceRequests (Maybe Int)

The maximum number of results to return in a single call. Specify a value between 5 and 1000. To retrieve the remaining results, make another call with the returned NextToken value.

describeSpotInstanceRequests_dryRun :: Lens' DescribeSpotInstanceRequests (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeSpotInstanceRequests_spotInstanceRequestIds :: Lens' DescribeSpotInstanceRequests (Maybe [Text])

One or more Spot Instance request IDs.

describeSpotInstanceRequests_filters :: Lens' DescribeSpotInstanceRequests (Maybe [Filter])

One or more filters.

availability-zone-group - The Availability Zone group.

create-time - The time stamp when the Spot Instance request was created.

fault-code - The fault code related to the request.

fault-message - The fault message related to the request.

instance-id - The ID of the instance that fulfilled the request.

launch-group - The Spot Instance launch group.

launch.block-device-mapping.delete-on-termination - Indicates whether the EBS volume is deleted on instance termination.

launch.block-device-mapping.device-name - The device name for the volume in the block device mapping (for example, /dev/sdh or xvdh).

launch.block-device-mapping.snapshot-id - The ID of the snapshot for the EBS volume.

launch.block-device-mapping.volume-size - The size of the EBS volume, in GiB.

launch.block-device-mapping.volume-type - The type of EBS volume: gp2 for General Purpose SSD, io1 or io2 for Provisioned IOPS SSD, st1 for Throughput Optimized HDD, sc1for Cold HDD, or standard for Magnetic.

launch.group-id - The ID of the security group for the instance.

launch.group-name - The name of the security group for the instance.

launch.image-id - The ID of the AMI.

launch.instance-type - The type of instance (for example, m3.medium).

launch.kernel-id - The kernel ID.

launch.key-name - The name of the key pair the instance launched with.

launch.monitoring-enabled - Whether detailed monitoring is enabled for the Spot Instance.

launch.ramdisk-id - The RAM disk ID.

launched-availability-zone - The Availability Zone in which the request is launched.

network-interface.addresses.primary - Indicates whether the IP address is the primary private IP address.

network-interface.delete-on-termination - Indicates whether the network interface is deleted when the instance is terminated.

network-interface.description - A description of the network interface.

network-interface.device-index - The index of the device for the network interface attachment on the instance.

network-interface.group-id - The ID of the security group associated with the network interface.

network-interface.network-interface-id - The ID of the network interface.

network-interface.private-ip-address - The primary private IP address of the network interface.

network-interface.subnet-id - The ID of the subnet for the instance.

product-description - The product description associated with the instance (Linux/UNIX | Windows).

spot-instance-request-id - The Spot Instance request ID.

spot-price - The maximum hourly price for any Spot Instance launched to fulfill the request.

state - The state of the Spot Instance request (open | active | closed | cancelled | failed). Spot request status information can help you track your Amazon EC2 Spot Instance requests. For more information, see Spot request status in the Amazon EC2 User Guide for Linux Instances.

status-code - The short code describing the most recent evaluation of your Spot Instance request.

status-message - The message explaining the status of the Spot Instance request.

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

type - The type of Spot Instance request (one-time | persistent).

valid-from - The start date of the request.

valid-until - The end date of the request.

Destructuring the Response

data DescribeSpotInstanceRequestsResponse

Contains the output of DescribeSpotInstanceRequests.

See: newDescribeSpotInstanceRequestsResponse smart constructor.

Constructors

DescribeSpotInstanceRequestsResponse'

Fields

nextToken :: Maybe Text

The token to use to retrieve the next set of results. This value is null when there are no more results to return.

spotInstanceRequests :: Maybe [SpotInstanceRequest]

One or more Spot Instance requests.

httpStatus :: Int

The response's http status code.

newDescribeSpotInstanceRequestsResponse

Arguments

:: Int
$sel:httpStatus:DescribeSpotInstanceRequestsResponse'

-> DescribeSpotInstanceRequestsResponse

Create a value of DescribeSpotInstanceRequestsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeSpotInstanceRequests', describeSpotInstanceRequestsResponse_nextToken - The token to use to retrieve the next set of results. This value is null when there are no more results to return.

$sel:spotInstanceRequests:DescribeSpotInstanceRequestsResponse', describeSpotInstanceRequestsResponse_spotInstanceRequests - One or more Spot Instance requests.

$sel:httpStatus:DescribeSpotInstanceRequestsResponse', describeSpotInstanceRequestsResponse_httpStatus - The response's http status code.

Response Lenses

describeSpotInstanceRequestsResponse_nextToken :: Lens' DescribeSpotInstanceRequestsResponse (Maybe Text)

The token to use to retrieve the next set of results. This value is null when there are no more results to return.

describeSpotInstanceRequestsResponse_spotInstanceRequests :: Lens' DescribeSpotInstanceRequestsResponse (Maybe [SpotInstanceRequest])

One or more Spot Instance requests.

describeSpotInstanceRequestsResponse_httpStatus :: Lens' DescribeSpotInstanceRequestsResponse Int

The response's http status code.