Network.AWS.EC2.DescribeSpotPriceHistory

Description

Describes the Spot price history. For more information, see Spot Instance pricing history in the Amazon EC2 User Guide for Linux Instances.

When you specify a start and end time, the operation returns the prices of the instance types within that time range. It also returns the last price change before the start time, which is the effective price as of the start time.

This operation returns paginated results.

Creating a Request

data DescribeSpotPriceHistory

Contains the parameters for DescribeSpotPriceHistory.

See: newDescribeSpotPriceHistory smart constructor.

Constructors

DescribeSpotPriceHistory'

Fields

nextToken :: Maybe Text

The token for the next set of results.

maxResults :: Maybe Int

The maximum number of results to return in a single call. Specify a value between 1 and 1000. The default value is 1000. To retrieve the remaining results, make another call with the returned NextToken value.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

startTime :: Maybe ISO8601

The date and time, up to the past 90 days, from which to start retrieving the price history data, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ).

endTime :: Maybe ISO8601

The date and time, up to the current date, from which to stop retrieving the price history data, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ).

availabilityZone :: Maybe Text

Filters the results by the specified Availability Zone.

productDescriptions :: Maybe [Text]

Filters the results by the specified basic product descriptions.

filters :: Maybe [Filter]

One or more filters.

availability-zone - The Availability Zone for which prices should be returned.

instance-type - The type of instance (for example, m3.medium).

product-description - The product description for the Spot price (Linux/UNIX | Red Hat Enterprise Linux | SUSE Linux | Windows | Linux/UNIX (Amazon VPC) | Red Hat Enterprise Linux (Amazon VPC) | SUSE Linux (Amazon VPC) | Windows (Amazon VPC)).

spot-price - The Spot price. The value must match exactly (or use wildcards; greater than or less than comparison is not supported).

timestamp - The time stamp of the Spot price history, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ). You can use wildcards (* and ?). Greater than or less than comparison is not supported.

instanceTypes :: Maybe [InstanceType]

Filters the results by the specified instance types.

newDescribeSpotPriceHistory :: DescribeSpotPriceHistory

Create a value of DescribeSpotPriceHistory with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeSpotPriceHistory', describeSpotPriceHistory_nextToken - The token for the next set of results.

$sel:maxResults:DescribeSpotPriceHistory', describeSpotPriceHistory_maxResults - The maximum number of results to return in a single call. Specify a value between 1 and 1000. The default value is 1000. To retrieve the remaining results, make another call with the returned NextToken value.

$sel:dryRun:DescribeSpotPriceHistory', describeSpotPriceHistory_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:startTime:DescribeSpotPriceHistory', describeSpotPriceHistory_startTime - The date and time, up to the past 90 days, from which to start retrieving the price history data, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ).

$sel:endTime:DescribeSpotPriceHistory', describeSpotPriceHistory_endTime - The date and time, up to the current date, from which to stop retrieving the price history data, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ).

$sel:availabilityZone:DescribeSpotPriceHistory', describeSpotPriceHistory_availabilityZone - Filters the results by the specified Availability Zone.

$sel:productDescriptions:DescribeSpotPriceHistory', describeSpotPriceHistory_productDescriptions - Filters the results by the specified basic product descriptions.

$sel:filters:DescribeSpotPriceHistory', describeSpotPriceHistory_filters - One or more filters.

availability-zone - The Availability Zone for which prices should be returned.

instance-type - The type of instance (for example, m3.medium).

product-description - The product description for the Spot price (Linux/UNIX | Red Hat Enterprise Linux | SUSE Linux | Windows | Linux/UNIX (Amazon VPC) | Red Hat Enterprise Linux (Amazon VPC) | SUSE Linux (Amazon VPC) | Windows (Amazon VPC)).

spot-price - The Spot price. The value must match exactly (or use wildcards; greater than or less than comparison is not supported).

timestamp - The time stamp of the Spot price history, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ). You can use wildcards (* and ?). Greater than or less than comparison is not supported.

$sel:instanceTypes:DescribeSpotPriceHistory', describeSpotPriceHistory_instanceTypes - Filters the results by the specified instance types.

Request Lenses

describeSpotPriceHistory_nextToken :: Lens' DescribeSpotPriceHistory (Maybe Text)

The token for the next set of results.

describeSpotPriceHistory_maxResults :: Lens' DescribeSpotPriceHistory (Maybe Int)

The maximum number of results to return in a single call. Specify a value between 1 and 1000. The default value is 1000. To retrieve the remaining results, make another call with the returned NextToken value.

describeSpotPriceHistory_dryRun :: Lens' DescribeSpotPriceHistory (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeSpotPriceHistory_startTime :: Lens' DescribeSpotPriceHistory (Maybe UTCTime)

The date and time, up to the past 90 days, from which to start retrieving the price history data, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ).

describeSpotPriceHistory_endTime :: Lens' DescribeSpotPriceHistory (Maybe UTCTime)

The date and time, up to the current date, from which to stop retrieving the price history data, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ).

describeSpotPriceHistory_availabilityZone :: Lens' DescribeSpotPriceHistory (Maybe Text)

Filters the results by the specified Availability Zone.

describeSpotPriceHistory_productDescriptions :: Lens' DescribeSpotPriceHistory (Maybe [Text])

Filters the results by the specified basic product descriptions.

describeSpotPriceHistory_filters :: Lens' DescribeSpotPriceHistory (Maybe [Filter])

One or more filters.

availability-zone - The Availability Zone for which prices should be returned.

instance-type - The type of instance (for example, m3.medium).

product-description - The product description for the Spot price (Linux/UNIX | Red Hat Enterprise Linux | SUSE Linux | Windows | Linux/UNIX (Amazon VPC) | Red Hat Enterprise Linux (Amazon VPC) | SUSE Linux (Amazon VPC) | Windows (Amazon VPC)).

spot-price - The Spot price. The value must match exactly (or use wildcards; greater than or less than comparison is not supported).

timestamp - The time stamp of the Spot price history, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ). You can use wildcards (* and ?). Greater than or less than comparison is not supported.

describeSpotPriceHistory_instanceTypes :: Lens' DescribeSpotPriceHistory (Maybe [InstanceType])

Filters the results by the specified instance types.

Destructuring the Response

data DescribeSpotPriceHistoryResponse

Contains the output of DescribeSpotPriceHistory.

See: newDescribeSpotPriceHistoryResponse smart constructor.

Constructors

DescribeSpotPriceHistoryResponse'

Fields

nextToken :: Maybe Text

The token required to retrieve the next set of results. This value is null or an empty string when there are no more results to return.

spotPriceHistory :: Maybe [SpotPrice]

The historical Spot prices.

httpStatus :: Int

The response's http status code.

newDescribeSpotPriceHistoryResponse

Arguments

:: Int
$sel:httpStatus:DescribeSpotPriceHistoryResponse'

-> DescribeSpotPriceHistoryResponse

Create a value of DescribeSpotPriceHistoryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeSpotPriceHistory', describeSpotPriceHistoryResponse_nextToken - The token required to retrieve the next set of results. This value is null or an empty string when there are no more results to return.

$sel:spotPriceHistory:DescribeSpotPriceHistoryResponse', describeSpotPriceHistoryResponse_spotPriceHistory - The historical Spot prices.

$sel:httpStatus:DescribeSpotPriceHistoryResponse', describeSpotPriceHistoryResponse_httpStatus - The response's http status code.

Response Lenses

describeSpotPriceHistoryResponse_nextToken :: Lens' DescribeSpotPriceHistoryResponse (Maybe Text)

The token required to retrieve the next set of results. This value is null or an empty string when there are no more results to return.

describeSpotPriceHistoryResponse_spotPriceHistory :: Lens' DescribeSpotPriceHistoryResponse (Maybe [SpotPrice])

The historical Spot prices.

describeSpotPriceHistoryResponse_httpStatus :: Lens' DescribeSpotPriceHistoryResponse Int

The response's http status code.