Network.AWS.EC2.DescribeStaleSecurityGroups

Description

VPC only

Describes the stale security group rules for security groups in a specified VPC. Rules are stale when they reference a deleted security group in a peer VPC, or a security group in a peer VPC for which the VPC peering connection has been deleted.

This operation returns paginated results.

Creating a Request

data DescribeStaleSecurityGroups

See: newDescribeStaleSecurityGroups smart constructor.

Constructors

DescribeStaleSecurityGroups'

Fields

nextToken :: Maybe Text

The token for the next set of items to return. (You received this token from a prior call.)

maxResults :: Maybe Natural

The maximum number of items to return for this request. The request returns a token that you can specify in a subsequent call to get the next set of results.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

vpcId :: Text

The ID of the VPC.

newDescribeStaleSecurityGroups

Arguments

:: Text
$sel:vpcId:DescribeStaleSecurityGroups'

-> DescribeStaleSecurityGroups

Create a value of DescribeStaleSecurityGroups with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeStaleSecurityGroups', describeStaleSecurityGroups_nextToken - The token for the next set of items to return. (You received this token from a prior call.)

$sel:maxResults:DescribeStaleSecurityGroups', describeStaleSecurityGroups_maxResults - The maximum number of items to return for this request. The request returns a token that you can specify in a subsequent call to get the next set of results.

$sel:dryRun:DescribeStaleSecurityGroups', describeStaleSecurityGroups_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:vpcId:DescribeStaleSecurityGroups', describeStaleSecurityGroups_vpcId - The ID of the VPC.

Request Lenses

describeStaleSecurityGroups_nextToken :: Lens' DescribeStaleSecurityGroups (Maybe Text)

The token for the next set of items to return. (You received this token from a prior call.)

describeStaleSecurityGroups_maxResults :: Lens' DescribeStaleSecurityGroups (Maybe Natural)

The maximum number of items to return for this request. The request returns a token that you can specify in a subsequent call to get the next set of results.

describeStaleSecurityGroups_dryRun :: Lens' DescribeStaleSecurityGroups (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeStaleSecurityGroups_vpcId :: Lens' DescribeStaleSecurityGroups Text

The ID of the VPC.

Destructuring the Response

data DescribeStaleSecurityGroupsResponse

See: newDescribeStaleSecurityGroupsResponse smart constructor.

Constructors

DescribeStaleSecurityGroupsResponse'

Fields

nextToken :: Maybe Text

The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

staleSecurityGroupSet :: Maybe [StaleSecurityGroup]

Information about the stale security groups.

httpStatus :: Int

The response's http status code.

newDescribeStaleSecurityGroupsResponse

Arguments

:: Int
$sel:httpStatus:DescribeStaleSecurityGroupsResponse'

-> DescribeStaleSecurityGroupsResponse

Create a value of DescribeStaleSecurityGroupsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeStaleSecurityGroups', describeStaleSecurityGroupsResponse_nextToken - The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

$sel:staleSecurityGroupSet:DescribeStaleSecurityGroupsResponse', describeStaleSecurityGroupsResponse_staleSecurityGroupSet - Information about the stale security groups.

$sel:httpStatus:DescribeStaleSecurityGroupsResponse', describeStaleSecurityGroupsResponse_httpStatus - The response's http status code.

Response Lenses

describeStaleSecurityGroupsResponse_nextToken :: Lens' DescribeStaleSecurityGroupsResponse (Maybe Text)

The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

describeStaleSecurityGroupsResponse_staleSecurityGroupSet :: Lens' DescribeStaleSecurityGroupsResponse (Maybe [StaleSecurityGroup])

Information about the stale security groups.

describeStaleSecurityGroupsResponse_httpStatus :: Lens' DescribeStaleSecurityGroupsResponse Int

The response's http status code.