Network.AWS.EC2.DescribeStoreImageTasks

Description

Describes the progress of the AMI store tasks. You can describe the store tasks for specified AMIs. If you don't specify the AMIs, you get a paginated list of store tasks from the last 31 days.

For each AMI task, the response indicates if the task is InProgress, Completed, or Failed. For tasks InProgress, the response shows the estimated progress as a percentage.

Tasks are listed in reverse chronological order. Currently, only tasks from the past 31 days can be viewed.

To use this API, you must have the required permissions. For more information, see Permissions for storing and restoring AMIs using Amazon S3 in the Amazon Elastic Compute Cloud User Guide.

For more information, see Store and restore an AMI using Amazon S3 in the Amazon Elastic Compute Cloud User Guide.

This operation returns paginated results.

Creating a Request

data DescribeStoreImageTasks

See: newDescribeStoreImageTasks smart constructor.

Constructors

DescribeStoreImageTasks'

Fields

filters :: Maybe [Filter]

The filters.

task-state - Returns tasks in a certain state (InProgress | Completed | Failed)

bucket - Returns task information for tasks that targeted a specific bucket. For the filter value, specify the bucket name.

imageIds :: Maybe [Text]

The AMI IDs for which to show progress. Up to 20 AMI IDs can be included in a request.

nextToken :: Maybe Text

The token for the next page of results.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

maxResults :: Maybe Natural

The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned NextToken value. This value can be between 1 and 200. You cannot specify this parameter and the ImageIDs parameter in the same call.

newDescribeStoreImageTasks :: DescribeStoreImageTasks

Create a value of DescribeStoreImageTasks with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:DescribeStoreImageTasks', describeStoreImageTasks_filters - The filters.

task-state - Returns tasks in a certain state (InProgress | Completed | Failed)

bucket - Returns task information for tasks that targeted a specific bucket. For the filter value, specify the bucket name.

$sel:imageIds:DescribeStoreImageTasks', describeStoreImageTasks_imageIds - The AMI IDs for which to show progress. Up to 20 AMI IDs can be included in a request.

$sel:nextToken:DescribeStoreImageTasks', describeStoreImageTasks_nextToken - The token for the next page of results.

$sel:dryRun:DescribeStoreImageTasks', describeStoreImageTasks_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:maxResults:DescribeStoreImageTasks', describeStoreImageTasks_maxResults - The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned NextToken value. This value can be between 1 and 200. You cannot specify this parameter and the ImageIDs parameter in the same call.

Request Lenses

describeStoreImageTasks_filters :: Lens' DescribeStoreImageTasks (Maybe [Filter])

The filters.

task-state - Returns tasks in a certain state (InProgress | Completed | Failed)

bucket - Returns task information for tasks that targeted a specific bucket. For the filter value, specify the bucket name.

describeStoreImageTasks_imageIds :: Lens' DescribeStoreImageTasks (Maybe [Text])

The AMI IDs for which to show progress. Up to 20 AMI IDs can be included in a request.

describeStoreImageTasks_nextToken :: Lens' DescribeStoreImageTasks (Maybe Text)

The token for the next page of results.

describeStoreImageTasks_dryRun :: Lens' DescribeStoreImageTasks (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeStoreImageTasks_maxResults :: Lens' DescribeStoreImageTasks (Maybe Natural)

The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned NextToken value. This value can be between 1 and 200. You cannot specify this parameter and the ImageIDs parameter in the same call.

Destructuring the Response

data DescribeStoreImageTasksResponse

See: newDescribeStoreImageTasksResponse smart constructor.

Constructors

DescribeStoreImageTasksResponse'

Fields

storeImageTaskResults :: Maybe [StoreImageTaskResult]

The information about the AMI store tasks.

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

httpStatus :: Int

The response's http status code.

newDescribeStoreImageTasksResponse

Arguments

:: Int
$sel:httpStatus:DescribeStoreImageTasksResponse'

-> DescribeStoreImageTasksResponse

Create a value of DescribeStoreImageTasksResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:storeImageTaskResults:DescribeStoreImageTasksResponse', describeStoreImageTasksResponse_storeImageTaskResults - The information about the AMI store tasks.

$sel:nextToken:DescribeStoreImageTasks', describeStoreImageTasksResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:httpStatus:DescribeStoreImageTasksResponse', describeStoreImageTasksResponse_httpStatus - The response's http status code.

Response Lenses

describeStoreImageTasksResponse_storeImageTaskResults :: Lens' DescribeStoreImageTasksResponse (Maybe [StoreImageTaskResult])

The information about the AMI store tasks.

describeStoreImageTasksResponse_nextToken :: Lens' DescribeStoreImageTasksResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

describeStoreImageTasksResponse_httpStatus :: Lens' DescribeStoreImageTasksResponse Int

The response's http status code.