Network.AWS.EC2.DescribeSubnets

Description

Describes one or more of your subnets.

For more information, see Your VPC and subnets in the Amazon Virtual Private Cloud User Guide.

This operation returns paginated results.

Creating a Request

data DescribeSubnets

See: newDescribeSubnets smart constructor.

Constructors

DescribeSubnets'

Fields

nextToken :: Maybe Text

The token for the next page of results.

maxResults :: Maybe Natural

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

subnetIds :: Maybe [Text]

One or more subnet IDs.

Default: Describes all your subnets.

filters :: Maybe [Filter]

One or more filters.

availability-zone - The Availability Zone for the subnet. You can also use availabilityZone as the filter name.

availability-zone-id - The ID of the Availability Zone for the subnet. You can also use availabilityZoneId as the filter name.

available-ip-address-count - The number of IPv4 addresses in the subnet that are available.

cidr-block - The IPv4 CIDR block of the subnet. The CIDR block you specify must exactly match the subnet's CIDR block for information to be returned for the subnet. You can also use cidr or cidrBlock as the filter names.

default-for-az - Indicates whether this is the default subnet for the Availability Zone. You can also use defaultForAz as the filter name.

ipv6-cidr-block-association.ipv6-cidr-block - An IPv6 CIDR block associated with the subnet.

ipv6-cidr-block-association.association-id - An association ID for an IPv6 CIDR block associated with the subnet.

ipv6-cidr-block-association.state - The state of an IPv6 CIDR block associated with the subnet.

outpost-arn - The Amazon Resource Name (ARN) of the Outpost.

owner-id - The ID of the Amazon Web Services account that owns the subnet.

state - The state of the subnet (pending | available).

subnet-arn - The Amazon Resource Name (ARN) of the subnet.

subnet-id - The ID of the subnet.

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

vpc-id - The ID of the VPC for the subnet.

newDescribeSubnets :: DescribeSubnets

Create a value of DescribeSubnets with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeSubnets', describeSubnets_nextToken - The token for the next page of results.

$sel:maxResults:DescribeSubnets', describeSubnets_maxResults - The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

$sel:dryRun:DescribeSubnets', describeSubnets_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:subnetIds:DescribeSubnets', describeSubnets_subnetIds - One or more subnet IDs.

Default: Describes all your subnets.

$sel:filters:DescribeSubnets', describeSubnets_filters - One or more filters.

availability-zone - The Availability Zone for the subnet. You can also use availabilityZone as the filter name.

availability-zone-id - The ID of the Availability Zone for the subnet. You can also use availabilityZoneId as the filter name.

available-ip-address-count - The number of IPv4 addresses in the subnet that are available.

cidr-block - The IPv4 CIDR block of the subnet. The CIDR block you specify must exactly match the subnet's CIDR block for information to be returned for the subnet. You can also use cidr or cidrBlock as the filter names.

default-for-az - Indicates whether this is the default subnet for the Availability Zone. You can also use defaultForAz as the filter name.

ipv6-cidr-block-association.ipv6-cidr-block - An IPv6 CIDR block associated with the subnet.

ipv6-cidr-block-association.association-id - An association ID for an IPv6 CIDR block associated with the subnet.

ipv6-cidr-block-association.state - The state of an IPv6 CIDR block associated with the subnet.

outpost-arn - The Amazon Resource Name (ARN) of the Outpost.

owner-id - The ID of the Amazon Web Services account that owns the subnet.

state - The state of the subnet (pending | available).

subnet-arn - The Amazon Resource Name (ARN) of the subnet.

subnet-id - The ID of the subnet.

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

vpc-id - The ID of the VPC for the subnet.

Request Lenses

describeSubnets_nextToken :: Lens' DescribeSubnets (Maybe Text)

The token for the next page of results.

describeSubnets_maxResults :: Lens' DescribeSubnets (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

describeSubnets_dryRun :: Lens' DescribeSubnets (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeSubnets_subnetIds :: Lens' DescribeSubnets (Maybe [Text])

One or more subnet IDs.

Default: Describes all your subnets.

describeSubnets_filters :: Lens' DescribeSubnets (Maybe [Filter])

One or more filters.

availability-zone - The Availability Zone for the subnet. You can also use availabilityZone as the filter name.

availability-zone-id - The ID of the Availability Zone for the subnet. You can also use availabilityZoneId as the filter name.

available-ip-address-count - The number of IPv4 addresses in the subnet that are available.

cidr-block - The IPv4 CIDR block of the subnet. The CIDR block you specify must exactly match the subnet's CIDR block for information to be returned for the subnet. You can also use cidr or cidrBlock as the filter names.

default-for-az - Indicates whether this is the default subnet for the Availability Zone. You can also use defaultForAz as the filter name.

ipv6-cidr-block-association.ipv6-cidr-block - An IPv6 CIDR block associated with the subnet.

ipv6-cidr-block-association.association-id - An association ID for an IPv6 CIDR block associated with the subnet.

ipv6-cidr-block-association.state - The state of an IPv6 CIDR block associated with the subnet.

outpost-arn - The Amazon Resource Name (ARN) of the Outpost.

owner-id - The ID of the Amazon Web Services account that owns the subnet.

state - The state of the subnet (pending | available).

subnet-arn - The Amazon Resource Name (ARN) of the subnet.

subnet-id - The ID of the subnet.

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

vpc-id - The ID of the VPC for the subnet.

Destructuring the Response

data DescribeSubnetsResponse

See: newDescribeSubnetsResponse smart constructor.

Constructors

DescribeSubnetsResponse'

Fields

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

subnets :: Maybe [Subnet]

Information about one or more subnets.

httpStatus :: Int

The response's http status code.

newDescribeSubnetsResponse

Arguments

:: Int
$sel:httpStatus:DescribeSubnetsResponse'

-> DescribeSubnetsResponse

Create a value of DescribeSubnetsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeSubnets', describeSubnetsResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:subnets:DescribeSubnetsResponse', describeSubnetsResponse_subnets - Information about one or more subnets.

$sel:httpStatus:DescribeSubnetsResponse', describeSubnetsResponse_httpStatus - The response's http status code.

Response Lenses

describeSubnetsResponse_nextToken :: Lens' DescribeSubnetsResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

describeSubnetsResponse_subnets :: Lens' DescribeSubnetsResponse (Maybe [Subnet])

Information about one or more subnets.

describeSubnetsResponse_httpStatus :: Lens' DescribeSubnetsResponse Int

The response's http status code.