Network.AWS.EC2.DescribeTransitGatewayMulticastDomains

Description

Describes one or more transit gateway multicast domains.

This operation returns paginated results.

Creating a Request

data DescribeTransitGatewayMulticastDomains

See: newDescribeTransitGatewayMulticastDomains smart constructor.

Constructors

DescribeTransitGatewayMulticastDomains'

Fields

nextToken :: Maybe Text

The token for the next page of results.

maxResults :: Maybe Natural

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

filters :: Maybe [Filter]

One or more filters. The possible values are:

state - The state of the transit gateway multicast domain. Valid values are pending | available | deleting | deleted.

transit-gateway-id - The ID of the transit gateway.

transit-gateway-multicast-domain-id - The ID of the transit gateway multicast domain.

transitGatewayMulticastDomainIds :: Maybe [Text]

The ID of the transit gateway multicast domain.

newDescribeTransitGatewayMulticastDomains :: DescribeTransitGatewayMulticastDomains

Create a value of DescribeTransitGatewayMulticastDomains with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeTransitGatewayMulticastDomains', describeTransitGatewayMulticastDomains_nextToken - The token for the next page of results.

$sel:maxResults:DescribeTransitGatewayMulticastDomains', describeTransitGatewayMulticastDomains_maxResults - The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

$sel:dryRun:DescribeTransitGatewayMulticastDomains', describeTransitGatewayMulticastDomains_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:filters:DescribeTransitGatewayMulticastDomains', describeTransitGatewayMulticastDomains_filters - One or more filters. The possible values are:

state - The state of the transit gateway multicast domain. Valid values are pending | available | deleting | deleted.

transit-gateway-id - The ID of the transit gateway.

transit-gateway-multicast-domain-id - The ID of the transit gateway multicast domain.

$sel:transitGatewayMulticastDomainIds:DescribeTransitGatewayMulticastDomains', describeTransitGatewayMulticastDomains_transitGatewayMulticastDomainIds - The ID of the transit gateway multicast domain.

Request Lenses

describeTransitGatewayMulticastDomains_nextToken :: Lens' DescribeTransitGatewayMulticastDomains (Maybe Text)

The token for the next page of results.

describeTransitGatewayMulticastDomains_maxResults :: Lens' DescribeTransitGatewayMulticastDomains (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

describeTransitGatewayMulticastDomains_dryRun :: Lens' DescribeTransitGatewayMulticastDomains (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeTransitGatewayMulticastDomains_filters :: Lens' DescribeTransitGatewayMulticastDomains (Maybe [Filter])

One or more filters. The possible values are:

state - The state of the transit gateway multicast domain. Valid values are pending | available | deleting | deleted.

transit-gateway-id - The ID of the transit gateway.

transit-gateway-multicast-domain-id - The ID of the transit gateway multicast domain.

describeTransitGatewayMulticastDomains_transitGatewayMulticastDomainIds :: Lens' DescribeTransitGatewayMulticastDomains (Maybe [Text])

The ID of the transit gateway multicast domain.

Destructuring the Response

data DescribeTransitGatewayMulticastDomainsResponse

See: newDescribeTransitGatewayMulticastDomainsResponse smart constructor.

Constructors

DescribeTransitGatewayMulticastDomainsResponse'

Fields

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

transitGatewayMulticastDomains :: Maybe [TransitGatewayMulticastDomain]

Information about the transit gateway multicast domains.

httpStatus :: Int

The response's http status code.

newDescribeTransitGatewayMulticastDomainsResponse

Arguments

:: Int
$sel:httpStatus:DescribeTransitGatewayMulticastDomainsResponse'

-> DescribeTransitGatewayMulticastDomainsResponse

Create a value of DescribeTransitGatewayMulticastDomainsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeTransitGatewayMulticastDomains', describeTransitGatewayMulticastDomainsResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:transitGatewayMulticastDomains:DescribeTransitGatewayMulticastDomainsResponse', describeTransitGatewayMulticastDomainsResponse_transitGatewayMulticastDomains - Information about the transit gateway multicast domains.

$sel:httpStatus:DescribeTransitGatewayMulticastDomainsResponse', describeTransitGatewayMulticastDomainsResponse_httpStatus - The response's http status code.

Response Lenses

describeTransitGatewayMulticastDomainsResponse_nextToken :: Lens' DescribeTransitGatewayMulticastDomainsResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

describeTransitGatewayMulticastDomainsResponse_transitGatewayMulticastDomains :: Lens' DescribeTransitGatewayMulticastDomainsResponse (Maybe [TransitGatewayMulticastDomain])

Information about the transit gateway multicast domains.

describeTransitGatewayMulticastDomainsResponse_httpStatus :: Lens' DescribeTransitGatewayMulticastDomainsResponse Int

The response's http status code.