Network.AWS.EC2.DescribeTransitGatewayVpcAttachments

Description

Describes one or more VPC attachments. By default, all VPC attachments are described. Alternatively, you can filter the results.

This operation returns paginated results.

Creating a Request

data DescribeTransitGatewayVpcAttachments

See: newDescribeTransitGatewayVpcAttachments smart constructor.

Constructors

DescribeTransitGatewayVpcAttachments'

Fields

filters :: Maybe [Filter]

One or more filters. The possible values are:

state - The state of the attachment. Valid values are available | deleted | deleting | failed | failing | initiatingRequest | modifying | pendingAcceptance | pending | rollingBack | rejected | rejecting.

transit-gateway-attachment-id - The ID of the attachment.

transit-gateway-id - The ID of the transit gateway.

vpc-id - The ID of the VPC.

nextToken :: Maybe Text

The token for the next page of results.

transitGatewayAttachmentIds :: Maybe [Text]

The IDs of the attachments.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

maxResults :: Maybe Natural

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

newDescribeTransitGatewayVpcAttachments :: DescribeTransitGatewayVpcAttachments

Create a value of DescribeTransitGatewayVpcAttachments with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:DescribeTransitGatewayVpcAttachments', describeTransitGatewayVpcAttachments_filters - One or more filters. The possible values are:

state - The state of the attachment. Valid values are available | deleted | deleting | failed | failing | initiatingRequest | modifying | pendingAcceptance | pending | rollingBack | rejected | rejecting.

transit-gateway-attachment-id - The ID of the attachment.

transit-gateway-id - The ID of the transit gateway.

vpc-id - The ID of the VPC.

$sel:nextToken:DescribeTransitGatewayVpcAttachments', describeTransitGatewayVpcAttachments_nextToken - The token for the next page of results.

$sel:transitGatewayAttachmentIds:DescribeTransitGatewayVpcAttachments', describeTransitGatewayVpcAttachments_transitGatewayAttachmentIds - The IDs of the attachments.

$sel:dryRun:DescribeTransitGatewayVpcAttachments', describeTransitGatewayVpcAttachments_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:maxResults:DescribeTransitGatewayVpcAttachments', describeTransitGatewayVpcAttachments_maxResults - The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

Request Lenses

describeTransitGatewayVpcAttachments_filters :: Lens' DescribeTransitGatewayVpcAttachments (Maybe [Filter])

One or more filters. The possible values are:

state - The state of the attachment. Valid values are available | deleted | deleting | failed | failing | initiatingRequest | modifying | pendingAcceptance | pending | rollingBack | rejected | rejecting.

transit-gateway-attachment-id - The ID of the attachment.

transit-gateway-id - The ID of the transit gateway.

vpc-id - The ID of the VPC.

describeTransitGatewayVpcAttachments_nextToken :: Lens' DescribeTransitGatewayVpcAttachments (Maybe Text)

The token for the next page of results.

describeTransitGatewayVpcAttachments_transitGatewayAttachmentIds :: Lens' DescribeTransitGatewayVpcAttachments (Maybe [Text])

The IDs of the attachments.

describeTransitGatewayVpcAttachments_dryRun :: Lens' DescribeTransitGatewayVpcAttachments (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeTransitGatewayVpcAttachments_maxResults :: Lens' DescribeTransitGatewayVpcAttachments (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

Destructuring the Response

data DescribeTransitGatewayVpcAttachmentsResponse

See: newDescribeTransitGatewayVpcAttachmentsResponse smart constructor.

Constructors

DescribeTransitGatewayVpcAttachmentsResponse'

Fields

transitGatewayVpcAttachments :: Maybe [TransitGatewayVpcAttachment]

Information about the VPC attachments.

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

httpStatus :: Int

The response's http status code.

newDescribeTransitGatewayVpcAttachmentsResponse

Arguments

:: Int
$sel:httpStatus:DescribeTransitGatewayVpcAttachmentsResponse'

-> DescribeTransitGatewayVpcAttachmentsResponse

Create a value of DescribeTransitGatewayVpcAttachmentsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:transitGatewayVpcAttachments:DescribeTransitGatewayVpcAttachmentsResponse', describeTransitGatewayVpcAttachmentsResponse_transitGatewayVpcAttachments - Information about the VPC attachments.

$sel:nextToken:DescribeTransitGatewayVpcAttachments', describeTransitGatewayVpcAttachmentsResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:httpStatus:DescribeTransitGatewayVpcAttachmentsResponse', describeTransitGatewayVpcAttachmentsResponse_httpStatus - The response's http status code.

Response Lenses

describeTransitGatewayVpcAttachmentsResponse_transitGatewayVpcAttachments :: Lens' DescribeTransitGatewayVpcAttachmentsResponse (Maybe [TransitGatewayVpcAttachment])

Information about the VPC attachments.

describeTransitGatewayVpcAttachmentsResponse_nextToken :: Lens' DescribeTransitGatewayVpcAttachmentsResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

describeTransitGatewayVpcAttachmentsResponse_httpStatus :: Lens' DescribeTransitGatewayVpcAttachmentsResponse Int

The response's http status code.