Network.AWS.EC2.DescribeVolumesModifications

Description

Describes the most recent volume modification request for the specified EBS volumes.

If a volume has never been modified, some information in the output will be null. If a volume has been modified more than once, the output includes only the most recent modification request.

You can also use CloudWatch Events to check the status of a modification to an EBS volume. For information about CloudWatch Events, see the Amazon CloudWatch Events User Guide. For more information, see Monitor the progress of volume modifications in the Amazon Elastic Compute Cloud User Guide.

This operation returns paginated results.

Creating a Request

data DescribeVolumesModifications

See: newDescribeVolumesModifications smart constructor.

Constructors

DescribeVolumesModifications'

Fields

nextToken :: Maybe Text

The nextToken value returned by a previous paginated request.

maxResults :: Maybe Int

The maximum number of results (up to a limit of 500) to be returned in a paginated request.

volumeIds :: Maybe [Text]

The IDs of the volumes.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

filters :: Maybe [Filter]

The filters.

modification-state - The current modification state (modifying | optimizing | completed | failed).

original-iops - The original IOPS rate of the volume.

original-size - The original size of the volume, in GiB.

original-volume-type - The original volume type of the volume (standard | io1 | io2 | gp2 | sc1 | st1).

originalMultiAttachEnabled - Indicates whether Multi-Attach support was enabled (true | false).

start-time - The modification start time.

target-iops - The target IOPS rate of the volume.

target-size - The target size of the volume, in GiB.

target-volume-type - The target volume type of the volume (standard | io1 | io2 | gp2 | sc1 | st1).

targetMultiAttachEnabled - Indicates whether Multi-Attach support is to be enabled (true | false).

volume-id - The ID of the volume.

newDescribeVolumesModifications :: DescribeVolumesModifications

Create a value of DescribeVolumesModifications with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeVolumesModifications', describeVolumesModifications_nextToken - The nextToken value returned by a previous paginated request.

$sel:maxResults:DescribeVolumesModifications', describeVolumesModifications_maxResults - The maximum number of results (up to a limit of 500) to be returned in a paginated request.

$sel:volumeIds:DescribeVolumesModifications', describeVolumesModifications_volumeIds - The IDs of the volumes.

$sel:dryRun:DescribeVolumesModifications', describeVolumesModifications_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:filters:DescribeVolumesModifications', describeVolumesModifications_filters - The filters.

modification-state - The current modification state (modifying | optimizing | completed | failed).

original-iops - The original IOPS rate of the volume.

original-size - The original size of the volume, in GiB.

original-volume-type - The original volume type of the volume (standard | io1 | io2 | gp2 | sc1 | st1).

originalMultiAttachEnabled - Indicates whether Multi-Attach support was enabled (true | false).

start-time - The modification start time.

target-iops - The target IOPS rate of the volume.

target-size - The target size of the volume, in GiB.

target-volume-type - The target volume type of the volume (standard | io1 | io2 | gp2 | sc1 | st1).

targetMultiAttachEnabled - Indicates whether Multi-Attach support is to be enabled (true | false).

volume-id - The ID of the volume.

Request Lenses

describeVolumesModifications_nextToken :: Lens' DescribeVolumesModifications (Maybe Text)

The nextToken value returned by a previous paginated request.

describeVolumesModifications_maxResults :: Lens' DescribeVolumesModifications (Maybe Int)

The maximum number of results (up to a limit of 500) to be returned in a paginated request.

describeVolumesModifications_volumeIds :: Lens' DescribeVolumesModifications (Maybe [Text])

The IDs of the volumes.

describeVolumesModifications_dryRun :: Lens' DescribeVolumesModifications (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeVolumesModifications_filters :: Lens' DescribeVolumesModifications (Maybe [Filter])

The filters.

modification-state - The current modification state (modifying | optimizing | completed | failed).

original-iops - The original IOPS rate of the volume.

original-size - The original size of the volume, in GiB.

original-volume-type - The original volume type of the volume (standard | io1 | io2 | gp2 | sc1 | st1).

originalMultiAttachEnabled - Indicates whether Multi-Attach support was enabled (true | false).

start-time - The modification start time.

target-iops - The target IOPS rate of the volume.

target-size - The target size of the volume, in GiB.

target-volume-type - The target volume type of the volume (standard | io1 | io2 | gp2 | sc1 | st1).

targetMultiAttachEnabled - Indicates whether Multi-Attach support is to be enabled (true | false).

volume-id - The ID of the volume.

Destructuring the Response

data DescribeVolumesModificationsResponse

See: newDescribeVolumesModificationsResponse smart constructor.

Constructors

DescribeVolumesModificationsResponse'

Fields

nextToken :: Maybe Text

Token for pagination, null if there are no more results

volumesModifications :: Maybe [VolumeModification]

Information about the volume modifications.

httpStatus :: Int

The response's http status code.

newDescribeVolumesModificationsResponse

Arguments

:: Int
$sel:httpStatus:DescribeVolumesModificationsResponse'

-> DescribeVolumesModificationsResponse

Create a value of DescribeVolumesModificationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeVolumesModifications', describeVolumesModificationsResponse_nextToken - Token for pagination, null if there are no more results

$sel:volumesModifications:DescribeVolumesModificationsResponse', describeVolumesModificationsResponse_volumesModifications - Information about the volume modifications.

$sel:httpStatus:DescribeVolumesModificationsResponse', describeVolumesModificationsResponse_httpStatus - The response's http status code.

Response Lenses

describeVolumesModificationsResponse_nextToken :: Lens' DescribeVolumesModificationsResponse (Maybe Text)

Token for pagination, null if there are no more results

describeVolumesModificationsResponse_volumesModifications :: Lens' DescribeVolumesModificationsResponse (Maybe [VolumeModification])

Information about the volume modifications.

describeVolumesModificationsResponse_httpStatus :: Lens' DescribeVolumesModificationsResponse Int

The response's http status code.