Network.AWS.EC2.DescribeVpcEndpointConnectionNotifications

Description

Describes the connection notifications for VPC endpoints and VPC endpoint services.

This operation returns paginated results.

Creating a Request

data DescribeVpcEndpointConnectionNotifications

See: newDescribeVpcEndpointConnectionNotifications smart constructor.

Constructors

DescribeVpcEndpointConnectionNotifications'

Fields

nextToken :: Maybe Text

The token to request the next page of results.

connectionNotificationId :: Maybe Text

The ID of the notification.

maxResults :: Maybe Int

The maximum number of results to return in a single call. To retrieve the remaining results, make another request with the returned NextToken value.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

filters :: Maybe [Filter]

One or more filters.

connection-notification-arn - The ARN of the SNS topic for the notification.

connection-notification-id - The ID of the notification.

connection-notification-state - The state of the notification (Enabled | Disabled).

connection-notification-type - The type of notification (Topic).

service-id - The ID of the endpoint service.

vpc-endpoint-id - The ID of the VPC endpoint.

newDescribeVpcEndpointConnectionNotifications :: DescribeVpcEndpointConnectionNotifications

Create a value of DescribeVpcEndpointConnectionNotifications with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeVpcEndpointConnectionNotifications', describeVpcEndpointConnectionNotifications_nextToken - The token to request the next page of results.

$sel:connectionNotificationId:DescribeVpcEndpointConnectionNotifications', describeVpcEndpointConnectionNotifications_connectionNotificationId - The ID of the notification.

$sel:maxResults:DescribeVpcEndpointConnectionNotifications', describeVpcEndpointConnectionNotifications_maxResults - The maximum number of results to return in a single call. To retrieve the remaining results, make another request with the returned NextToken value.

$sel:dryRun:DescribeVpcEndpointConnectionNotifications', describeVpcEndpointConnectionNotifications_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:filters:DescribeVpcEndpointConnectionNotifications', describeVpcEndpointConnectionNotifications_filters - One or more filters.

connection-notification-arn - The ARN of the SNS topic for the notification.

connection-notification-id - The ID of the notification.

connection-notification-state - The state of the notification (Enabled | Disabled).

connection-notification-type - The type of notification (Topic).

service-id - The ID of the endpoint service.

vpc-endpoint-id - The ID of the VPC endpoint.

Request Lenses

describeVpcEndpointConnectionNotifications_nextToken :: Lens' DescribeVpcEndpointConnectionNotifications (Maybe Text)

The token to request the next page of results.

describeVpcEndpointConnectionNotifications_connectionNotificationId :: Lens' DescribeVpcEndpointConnectionNotifications (Maybe Text)

The ID of the notification.

describeVpcEndpointConnectionNotifications_maxResults :: Lens' DescribeVpcEndpointConnectionNotifications (Maybe Int)

The maximum number of results to return in a single call. To retrieve the remaining results, make another request with the returned NextToken value.

describeVpcEndpointConnectionNotifications_dryRun :: Lens' DescribeVpcEndpointConnectionNotifications (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeVpcEndpointConnectionNotifications_filters :: Lens' DescribeVpcEndpointConnectionNotifications (Maybe [Filter])

One or more filters.

connection-notification-arn - The ARN of the SNS topic for the notification.

connection-notification-id - The ID of the notification.

connection-notification-state - The state of the notification (Enabled | Disabled).

connection-notification-type - The type of notification (Topic).

service-id - The ID of the endpoint service.

vpc-endpoint-id - The ID of the VPC endpoint.

Destructuring the Response

data DescribeVpcEndpointConnectionNotificationsResponse

See: newDescribeVpcEndpointConnectionNotificationsResponse smart constructor.

Constructors

DescribeVpcEndpointConnectionNotificationsResponse'

Fields

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

connectionNotificationSet :: Maybe [ConnectionNotification]

One or more notifications.

httpStatus :: Int

The response's http status code.

newDescribeVpcEndpointConnectionNotificationsResponse

Arguments

:: Int
$sel:httpStatus:DescribeVpcEndpointConnectionNotificationsResponse'

-> DescribeVpcEndpointConnectionNotificationsResponse

Create a value of DescribeVpcEndpointConnectionNotificationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeVpcEndpointConnectionNotifications', describeVpcEndpointConnectionNotificationsResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:connectionNotificationSet:DescribeVpcEndpointConnectionNotificationsResponse', describeVpcEndpointConnectionNotificationsResponse_connectionNotificationSet - One or more notifications.

$sel:httpStatus:DescribeVpcEndpointConnectionNotificationsResponse', describeVpcEndpointConnectionNotificationsResponse_httpStatus - The response's http status code.

Response Lenses

describeVpcEndpointConnectionNotificationsResponse_nextToken :: Lens' DescribeVpcEndpointConnectionNotificationsResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

describeVpcEndpointConnectionNotificationsResponse_connectionNotificationSet :: Lens' DescribeVpcEndpointConnectionNotificationsResponse (Maybe [ConnectionNotification])

One or more notifications.

describeVpcEndpointConnectionNotificationsResponse_httpStatus :: Lens' DescribeVpcEndpointConnectionNotificationsResponse Int

The response's http status code.