Network.AWS.EC2.DescribeVpcEndpointServices

Description

Describes available services to which you can create a VPC endpoint.

When the service provider and the consumer have different accounts in multiple Availability Zones, and the consumer views the VPC endpoint service information, the response only includes the common Availability Zones. For example, when the service provider account uses us-east-1a and us-east-1c and the consumer uses us-east-1a and us-east-1b, the response includes the VPC endpoint services in the common Availability Zone, us-east-1a.

This operation returns paginated results.

Creating a Request

data DescribeVpcEndpointServices

Contains the parameters for DescribeVpcEndpointServices.

See: newDescribeVpcEndpointServices smart constructor.

Constructors

DescribeVpcEndpointServices'

Fields

nextToken :: Maybe Text

The token for the next set of items to return. (You received this token from a prior call.)

serviceNames :: Maybe [Text]

One or more service names.

maxResults :: Maybe Int

The maximum number of items to return for this request. The request returns a token that you can specify in a subsequent call to get the next set of results.

Constraint: If the value is greater than 1,000, we return only 1,000 items.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

filters :: Maybe [Filter]

One or more filters.

service-name - The name of the service.

service-type - The type of service (Interface | Gateway).

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

newDescribeVpcEndpointServices :: DescribeVpcEndpointServices

Create a value of DescribeVpcEndpointServices with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeVpcEndpointServices', describeVpcEndpointServices_nextToken - The token for the next set of items to return. (You received this token from a prior call.)

$sel:serviceNames:DescribeVpcEndpointServices', describeVpcEndpointServices_serviceNames - One or more service names.

$sel:maxResults:DescribeVpcEndpointServices', describeVpcEndpointServices_maxResults - The maximum number of items to return for this request. The request returns a token that you can specify in a subsequent call to get the next set of results.

Constraint: If the value is greater than 1,000, we return only 1,000 items.

$sel:dryRun:DescribeVpcEndpointServices', describeVpcEndpointServices_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:filters:DescribeVpcEndpointServices', describeVpcEndpointServices_filters - One or more filters.

service-name - The name of the service.

service-type - The type of service (Interface | Gateway).

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

Request Lenses

describeVpcEndpointServices_nextToken :: Lens' DescribeVpcEndpointServices (Maybe Text)

The token for the next set of items to return. (You received this token from a prior call.)

describeVpcEndpointServices_serviceNames :: Lens' DescribeVpcEndpointServices (Maybe [Text])

One or more service names.

describeVpcEndpointServices_maxResults :: Lens' DescribeVpcEndpointServices (Maybe Int)

The maximum number of items to return for this request. The request returns a token that you can specify in a subsequent call to get the next set of results.

Constraint: If the value is greater than 1,000, we return only 1,000 items.

describeVpcEndpointServices_dryRun :: Lens' DescribeVpcEndpointServices (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeVpcEndpointServices_filters :: Lens' DescribeVpcEndpointServices (Maybe [Filter])

One or more filters.

service-name - The name of the service.

service-type - The type of service (Interface | Gateway).

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

Destructuring the Response

data DescribeVpcEndpointServicesResponse

Contains the output of DescribeVpcEndpointServices.

See: newDescribeVpcEndpointServicesResponse smart constructor.

Constructors

DescribeVpcEndpointServicesResponse'

Fields

serviceDetails :: Maybe [ServiceDetail]

Information about the service.

nextToken :: Maybe Text

The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

serviceNames :: Maybe [Text]

A list of supported services.

httpStatus :: Int

The response's http status code.

newDescribeVpcEndpointServicesResponse

Arguments

:: Int
$sel:httpStatus:DescribeVpcEndpointServicesResponse'

-> DescribeVpcEndpointServicesResponse

Create a value of DescribeVpcEndpointServicesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceDetails:DescribeVpcEndpointServicesResponse', describeVpcEndpointServicesResponse_serviceDetails - Information about the service.

$sel:nextToken:DescribeVpcEndpointServices', describeVpcEndpointServicesResponse_nextToken - The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

$sel:serviceNames:DescribeVpcEndpointServices', describeVpcEndpointServicesResponse_serviceNames - A list of supported services.

$sel:httpStatus:DescribeVpcEndpointServicesResponse', describeVpcEndpointServicesResponse_httpStatus - The response's http status code.

Response Lenses

describeVpcEndpointServicesResponse_serviceDetails :: Lens' DescribeVpcEndpointServicesResponse (Maybe [ServiceDetail])

Information about the service.

describeVpcEndpointServicesResponse_nextToken :: Lens' DescribeVpcEndpointServicesResponse (Maybe Text)

The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

describeVpcEndpointServicesResponse_serviceNames :: Lens' DescribeVpcEndpointServicesResponse (Maybe [Text])

A list of supported services.

describeVpcEndpointServicesResponse_httpStatus :: Lens' DescribeVpcEndpointServicesResponse Int

The response's http status code.