Network.AWS.EC2.DescribeVpcEndpoints

Description

Describes one or more of your VPC endpoints.

This operation returns paginated results.

Creating a Request

data DescribeVpcEndpoints

Contains the parameters for DescribeVpcEndpoints.

See: newDescribeVpcEndpoints smart constructor.

Constructors

DescribeVpcEndpoints'

Fields

nextToken :: Maybe Text

The token for the next set of items to return. (You received this token from a prior call.)

maxResults :: Maybe Int

The maximum number of items to return for this request. The request returns a token that you can specify in a subsequent call to get the next set of results.

Constraint: If the value is greater than 1,000, we return only 1,000 items.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

filters :: Maybe [Filter]

One or more filters.

service-name - The name of the service.

vpc-id - The ID of the VPC in which the endpoint resides.

vpc-endpoint-id - The ID of the endpoint.

vpc-endpoint-state - The state of the endpoint (pendingAcceptance | pending | available | deleting | deleted | rejected | failed).

vpc-endpoint-type - The type of VPC endpoint (Interface | Gateway | GatewayLoadBalancer).

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

vpcEndpointIds :: Maybe [Text]

One or more endpoint IDs.

newDescribeVpcEndpoints :: DescribeVpcEndpoints

Create a value of DescribeVpcEndpoints with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeVpcEndpoints', describeVpcEndpoints_nextToken - The token for the next set of items to return. (You received this token from a prior call.)

$sel:maxResults:DescribeVpcEndpoints', describeVpcEndpoints_maxResults - The maximum number of items to return for this request. The request returns a token that you can specify in a subsequent call to get the next set of results.

Constraint: If the value is greater than 1,000, we return only 1,000 items.

$sel:dryRun:DescribeVpcEndpoints', describeVpcEndpoints_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:filters:DescribeVpcEndpoints', describeVpcEndpoints_filters - One or more filters.

service-name - The name of the service.

vpc-id - The ID of the VPC in which the endpoint resides.

vpc-endpoint-id - The ID of the endpoint.

vpc-endpoint-state - The state of the endpoint (pendingAcceptance | pending | available | deleting | deleted | rejected | failed).

vpc-endpoint-type - The type of VPC endpoint (Interface | Gateway | GatewayLoadBalancer).

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

$sel:vpcEndpointIds:DescribeVpcEndpoints', describeVpcEndpoints_vpcEndpointIds - One or more endpoint IDs.

Request Lenses

describeVpcEndpoints_nextToken :: Lens' DescribeVpcEndpoints (Maybe Text)

The token for the next set of items to return. (You received this token from a prior call.)

describeVpcEndpoints_maxResults :: Lens' DescribeVpcEndpoints (Maybe Int)

The maximum number of items to return for this request. The request returns a token that you can specify in a subsequent call to get the next set of results.

Constraint: If the value is greater than 1,000, we return only 1,000 items.

describeVpcEndpoints_dryRun :: Lens' DescribeVpcEndpoints (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeVpcEndpoints_filters :: Lens' DescribeVpcEndpoints (Maybe [Filter])

One or more filters.

service-name - The name of the service.

vpc-id - The ID of the VPC in which the endpoint resides.

vpc-endpoint-id - The ID of the endpoint.

vpc-endpoint-state - The state of the endpoint (pendingAcceptance | pending | available | deleting | deleted | rejected | failed).

vpc-endpoint-type - The type of VPC endpoint (Interface | Gateway | GatewayLoadBalancer).

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

describeVpcEndpoints_vpcEndpointIds :: Lens' DescribeVpcEndpoints (Maybe [Text])

One or more endpoint IDs.

Destructuring the Response

data DescribeVpcEndpointsResponse

Contains the output of DescribeVpcEndpoints.

See: newDescribeVpcEndpointsResponse smart constructor.

Constructors

DescribeVpcEndpointsResponse'

Fields

nextToken :: Maybe Text

The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

vpcEndpoints :: Maybe [VpcEndpoint]

Information about the endpoints.

httpStatus :: Int

The response's http status code.

newDescribeVpcEndpointsResponse

Arguments

:: Int
$sel:httpStatus:DescribeVpcEndpointsResponse'

-> DescribeVpcEndpointsResponse

Create a value of DescribeVpcEndpointsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeVpcEndpoints', describeVpcEndpointsResponse_nextToken - The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

$sel:vpcEndpoints:DescribeVpcEndpointsResponse', describeVpcEndpointsResponse_vpcEndpoints - Information about the endpoints.

$sel:httpStatus:DescribeVpcEndpointsResponse', describeVpcEndpointsResponse_httpStatus - The response's http status code.

Response Lenses

describeVpcEndpointsResponse_nextToken :: Lens' DescribeVpcEndpointsResponse (Maybe Text)

The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

describeVpcEndpointsResponse_vpcEndpoints :: Lens' DescribeVpcEndpointsResponse (Maybe [VpcEndpoint])

Information about the endpoints.

describeVpcEndpointsResponse_httpStatus :: Lens' DescribeVpcEndpointsResponse Int

The response's http status code.