Network.AWS.EC2.DescribeVpnConnections

Description

Describes one or more of your VPN connections.

For more information, see Amazon Web Services Site-to-Site VPN in the Amazon Web Services Site-to-Site VPN User Guide.

Creating a Request

data DescribeVpnConnections

Contains the parameters for DescribeVpnConnections.

See: newDescribeVpnConnections smart constructor.

Constructors

DescribeVpnConnections'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

filters :: Maybe [Filter]

One or more filters.

customer-gateway-configuration - The configuration information for the customer gateway.

customer-gateway-id - The ID of a customer gateway associated with the VPN connection.

state - The state of the VPN connection (pending | available | deleting | deleted).

option.static-routes-only - Indicates whether the connection has static routes only. Used for devices that do not support Border Gateway Protocol (BGP).

route.destination-cidr-block - The destination CIDR block. This corresponds to the subnet used in a customer data center.

bgp-asn - The BGP Autonomous System Number (ASN) associated with a BGP device.

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

type - The type of VPN connection. Currently the only supported type is ipsec.1.

vpn-connection-id - The ID of the VPN connection.

vpn-gateway-id - The ID of a virtual private gateway associated with the VPN connection.

transit-gateway-id - The ID of a transit gateway associated with the VPN connection.

vpnConnectionIds :: Maybe [Text]

One or more VPN connection IDs.

Default: Describes your VPN connections.

newDescribeVpnConnections :: DescribeVpnConnections

Create a value of DescribeVpnConnections with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:DescribeVpnConnections', describeVpnConnections_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:filters:DescribeVpnConnections', describeVpnConnections_filters - One or more filters.

customer-gateway-configuration - The configuration information for the customer gateway.

customer-gateway-id - The ID of a customer gateway associated with the VPN connection.

state - The state of the VPN connection (pending | available | deleting | deleted).

option.static-routes-only - Indicates whether the connection has static routes only. Used for devices that do not support Border Gateway Protocol (BGP).

route.destination-cidr-block - The destination CIDR block. This corresponds to the subnet used in a customer data center.

bgp-asn - The BGP Autonomous System Number (ASN) associated with a BGP device.

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

type - The type of VPN connection. Currently the only supported type is ipsec.1.

vpn-connection-id - The ID of the VPN connection.

vpn-gateway-id - The ID of a virtual private gateway associated with the VPN connection.

transit-gateway-id - The ID of a transit gateway associated with the VPN connection.

$sel:vpnConnectionIds:DescribeVpnConnections', describeVpnConnections_vpnConnectionIds - One or more VPN connection IDs.

Default: Describes your VPN connections.

Request Lenses

describeVpnConnections_dryRun :: Lens' DescribeVpnConnections (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeVpnConnections_filters :: Lens' DescribeVpnConnections (Maybe [Filter])

One or more filters.

customer-gateway-configuration - The configuration information for the customer gateway.

customer-gateway-id - The ID of a customer gateway associated with the VPN connection.

state - The state of the VPN connection (pending | available | deleting | deleted).

option.static-routes-only - Indicates whether the connection has static routes only. Used for devices that do not support Border Gateway Protocol (BGP).

route.destination-cidr-block - The destination CIDR block. This corresponds to the subnet used in a customer data center.

bgp-asn - The BGP Autonomous System Number (ASN) associated with a BGP device.

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

type - The type of VPN connection. Currently the only supported type is ipsec.1.

vpn-connection-id - The ID of the VPN connection.

vpn-gateway-id - The ID of a virtual private gateway associated with the VPN connection.

transit-gateway-id - The ID of a transit gateway associated with the VPN connection.

describeVpnConnections_vpnConnectionIds :: Lens' DescribeVpnConnections (Maybe [Text])

One or more VPN connection IDs.

Default: Describes your VPN connections.

Destructuring the Response

data DescribeVpnConnectionsResponse

Contains the output of DescribeVpnConnections.

See: newDescribeVpnConnectionsResponse smart constructor.

Constructors

DescribeVpnConnectionsResponse'

Fields

vpnConnections :: Maybe [VpnConnection]

Information about one or more VPN connections.

httpStatus :: Int

The response's http status code.

newDescribeVpnConnectionsResponse

Arguments

:: Int
$sel:httpStatus:DescribeVpnConnectionsResponse'

-> DescribeVpnConnectionsResponse

Create a value of DescribeVpnConnectionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpnConnections:DescribeVpnConnectionsResponse', describeVpnConnectionsResponse_vpnConnections - Information about one or more VPN connections.

$sel:httpStatus:DescribeVpnConnectionsResponse', describeVpnConnectionsResponse_httpStatus - The response's http status code.

Response Lenses

describeVpnConnectionsResponse_vpnConnections :: Lens' DescribeVpnConnectionsResponse (Maybe [VpnConnection])

Information about one or more VPN connections.

describeVpnConnectionsResponse_httpStatus :: Lens' DescribeVpnConnectionsResponse Int

The response's http status code.