Network.AWS.EC2.DisassociateAddress

Description

Disassociates an Elastic IP address from the instance or network interface it's associated with.

An Elastic IP address is for use in either the EC2-Classic platform or in a VPC. For more information, see Elastic IP Addresses in the Amazon Elastic Compute Cloud User Guide.

This is an idempotent operation. If you perform the operation more than once, Amazon EC2 doesn't return an error.

Creating a Request

data DisassociateAddress

See: newDisassociateAddress smart constructor.

Constructors

DisassociateAddress'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

associationId :: Maybe Text

EC2-VPC

The association ID. Required for EC2-VPC.

publicIp :: Maybe Text

EC2-Classic

The Elastic IP address. Required for EC2-Classic.

newDisassociateAddress :: DisassociateAddress

Create a value of DisassociateAddress with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:DisassociateAddress', disassociateAddress_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:associationId:DisassociateAddress', disassociateAddress_associationId - [EC2-VPC] The association ID. Required for EC2-VPC.

$sel:publicIp:DisassociateAddress', disassociateAddress_publicIp - [EC2-Classic] The Elastic IP address. Required for EC2-Classic.

Request Lenses

disassociateAddress_dryRun :: Lens' DisassociateAddress (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

disassociateAddress_associationId :: Lens' DisassociateAddress (Maybe Text)

EC2-VPC

The association ID. Required for EC2-VPC.

disassociateAddress_publicIp :: Lens' DisassociateAddress (Maybe Text)

EC2-Classic

The Elastic IP address. Required for EC2-Classic.

Destructuring the Response

data DisassociateAddressResponse

See: newDisassociateAddressResponse smart constructor.

Constructors

DisassociateAddressResponse'

newDisassociateAddressResponse :: DisassociateAddressResponse

Create a value of DisassociateAddressResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.