Network.AWS.EC2.EnableFastSnapshotRestores

Description

Enables fast snapshot restores for the specified snapshots in the specified Availability Zones.

You get the full benefit of fast snapshot restores after they enter the enabled state. To get the current state of fast snapshot restores, use DescribeFastSnapshotRestores. To disable fast snapshot restores, use DisableFastSnapshotRestores.

For more information, see Amazon EBS fast snapshot restore in the Amazon Elastic Compute Cloud User Guide.

Creating a Request

data EnableFastSnapshotRestores

See: newEnableFastSnapshotRestores smart constructor.

Constructors

EnableFastSnapshotRestores'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

availabilityZones :: [Text]

One or more Availability Zones. For example, us-east-2a.

sourceSnapshotIds :: [Text]

The IDs of one or more snapshots. For example, snap-1234567890abcdef0. You can specify a snapshot that was shared with you from another Amazon Web Services account.

newEnableFastSnapshotRestores :: EnableFastSnapshotRestores

Create a value of EnableFastSnapshotRestores with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:EnableFastSnapshotRestores', enableFastSnapshotRestores_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:availabilityZones:EnableFastSnapshotRestores', enableFastSnapshotRestores_availabilityZones - One or more Availability Zones. For example, us-east-2a.

$sel:sourceSnapshotIds:EnableFastSnapshotRestores', enableFastSnapshotRestores_sourceSnapshotIds - The IDs of one or more snapshots. For example, snap-1234567890abcdef0. You can specify a snapshot that was shared with you from another Amazon Web Services account.

Request Lenses

enableFastSnapshotRestores_dryRun :: Lens' EnableFastSnapshotRestores (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

enableFastSnapshotRestores_availabilityZones :: Lens' EnableFastSnapshotRestores [Text]

One or more Availability Zones. For example, us-east-2a.

enableFastSnapshotRestores_sourceSnapshotIds :: Lens' EnableFastSnapshotRestores [Text]

The IDs of one or more snapshots. For example, snap-1234567890abcdef0. You can specify a snapshot that was shared with you from another Amazon Web Services account.

Destructuring the Response

data EnableFastSnapshotRestoresResponse

See: newEnableFastSnapshotRestoresResponse smart constructor.

Constructors

EnableFastSnapshotRestoresResponse'

Fields

unsuccessful :: Maybe [EnableFastSnapshotRestoreErrorItem]

Information about the snapshots for which fast snapshot restores could not be enabled.

successful :: Maybe [EnableFastSnapshotRestoreSuccessItem]

Information about the snapshots for which fast snapshot restores were successfully enabled.

httpStatus :: Int

The response's http status code.

newEnableFastSnapshotRestoresResponse

Arguments

:: Int
$sel:httpStatus:EnableFastSnapshotRestoresResponse'

-> EnableFastSnapshotRestoresResponse

Create a value of EnableFastSnapshotRestoresResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:unsuccessful:EnableFastSnapshotRestoresResponse', enableFastSnapshotRestoresResponse_unsuccessful - Information about the snapshots for which fast snapshot restores could not be enabled.

$sel:successful:EnableFastSnapshotRestoresResponse', enableFastSnapshotRestoresResponse_successful - Information about the snapshots for which fast snapshot restores were successfully enabled.

$sel:httpStatus:EnableFastSnapshotRestoresResponse', enableFastSnapshotRestoresResponse_httpStatus - The response's http status code.

Response Lenses

enableFastSnapshotRestoresResponse_unsuccessful :: Lens' EnableFastSnapshotRestoresResponse (Maybe [EnableFastSnapshotRestoreErrorItem])

Information about the snapshots for which fast snapshot restores could not be enabled.

enableFastSnapshotRestoresResponse_successful :: Lens' EnableFastSnapshotRestoresResponse (Maybe [EnableFastSnapshotRestoreSuccessItem])

Information about the snapshots for which fast snapshot restores were successfully enabled.

enableFastSnapshotRestoresResponse_httpStatus :: Lens' EnableFastSnapshotRestoresResponse Int

The response's http status code.