Network.AWS.EC2.EnableImageDeprecation

Description

Enables deprecation of the specified AMI at the specified date and time.

For more information, see Deprecate an AMI in the Amazon Elastic Compute Cloud User Guide.

Creating a Request

data EnableImageDeprecation

See: newEnableImageDeprecation smart constructor.

Constructors

EnableImageDeprecation'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

imageId :: Text

The ID of the AMI.

deprecateAt :: ISO8601

The date and time to deprecate the AMI, in UTC, in the following format: YYYY-MM-DDTHH:MM:SSZ. If you specify a value for seconds, Amazon EC2 rounds the seconds to the nearest minute.

You can’t specify a date in the past. The upper limit for DeprecateAt is 10 years from now.

newEnableImageDeprecation

Arguments

:: Text
$sel:imageId:EnableImageDeprecation'

-> UTCTime
$sel:deprecateAt:EnableImageDeprecation'

-> EnableImageDeprecation

Create a value of EnableImageDeprecation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:EnableImageDeprecation', enableImageDeprecation_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:imageId:EnableImageDeprecation', enableImageDeprecation_imageId - The ID of the AMI.

$sel:deprecateAt:EnableImageDeprecation', enableImageDeprecation_deprecateAt - The date and time to deprecate the AMI, in UTC, in the following format: YYYY-MM-DDTHH:MM:SSZ. If you specify a value for seconds, Amazon EC2 rounds the seconds to the nearest minute.

You can’t specify a date in the past. The upper limit for DeprecateAt is 10 years from now.

Request Lenses

enableImageDeprecation_dryRun :: Lens' EnableImageDeprecation (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

enableImageDeprecation_imageId :: Lens' EnableImageDeprecation Text

The ID of the AMI.

enableImageDeprecation_deprecateAt :: Lens' EnableImageDeprecation UTCTime

The date and time to deprecate the AMI, in UTC, in the following format: YYYY-MM-DDTHH:MM:SSZ. If you specify a value for seconds, Amazon EC2 rounds the seconds to the nearest minute.

You can’t specify a date in the past. The upper limit for DeprecateAt is 10 years from now.

Destructuring the Response

data EnableImageDeprecationResponse

See: newEnableImageDeprecationResponse smart constructor.

Constructors

EnableImageDeprecationResponse'

Fields

return' :: Maybe Bool

Returns true if the request succeeds; otherwise, it returns an error.

httpStatus :: Int

The response's http status code.

newEnableImageDeprecationResponse

Arguments

:: Int
$sel:httpStatus:EnableImageDeprecationResponse'

-> EnableImageDeprecationResponse

Create a value of EnableImageDeprecationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:return':EnableImageDeprecationResponse', enableImageDeprecationResponse_return - Returns true if the request succeeds; otherwise, it returns an error.

$sel:httpStatus:EnableImageDeprecationResponse', enableImageDeprecationResponse_httpStatus - The response's http status code.

Response Lenses

enableImageDeprecationResponse_return :: Lens' EnableImageDeprecationResponse (Maybe Bool)

Returns true if the request succeeds; otherwise, it returns an error.

enableImageDeprecationResponse_httpStatus :: Lens' EnableImageDeprecationResponse Int

The response's http status code.