Network.AWS.EC2.EnableVpcClassicLink

Description

Enables a VPC for ClassicLink. You can then link EC2-Classic instances to your ClassicLink-enabled VPC to allow communication over private IP addresses. You cannot enable your VPC for ClassicLink if any of your VPC route tables have existing routes for address ranges within the 10.0.0.0/8 IP address range, excluding local routes for VPCs in the 10.0.0.0/16 and 10.1.0.0/16 IP address ranges. For more information, see ClassicLink in the Amazon Elastic Compute Cloud User Guide.

Creating a Request

data EnableVpcClassicLink

See: newEnableVpcClassicLink smart constructor.

Constructors

EnableVpcClassicLink'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

vpcId :: Text

The ID of the VPC.

newEnableVpcClassicLink

Arguments

:: Text
$sel:vpcId:EnableVpcClassicLink'

-> EnableVpcClassicLink

Create a value of EnableVpcClassicLink with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:EnableVpcClassicLink', enableVpcClassicLink_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:vpcId:EnableVpcClassicLink', enableVpcClassicLink_vpcId - The ID of the VPC.

Request Lenses

enableVpcClassicLink_dryRun :: Lens' EnableVpcClassicLink (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

enableVpcClassicLink_vpcId :: Lens' EnableVpcClassicLink Text

The ID of the VPC.

Destructuring the Response

data EnableVpcClassicLinkResponse

See: newEnableVpcClassicLinkResponse smart constructor.

Constructors

EnableVpcClassicLinkResponse'

Fields

return' :: Maybe Bool

Returns true if the request succeeds; otherwise, it returns an error.

httpStatus :: Int

The response's http status code.

newEnableVpcClassicLinkResponse

Arguments

:: Int
$sel:httpStatus:EnableVpcClassicLinkResponse'

-> EnableVpcClassicLinkResponse

Create a value of EnableVpcClassicLinkResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:return':EnableVpcClassicLinkResponse', enableVpcClassicLinkResponse_return - Returns true if the request succeeds; otherwise, it returns an error.

$sel:httpStatus:EnableVpcClassicLinkResponse', enableVpcClassicLinkResponse_httpStatus - The response's http status code.

Response Lenses

enableVpcClassicLinkResponse_return :: Lens' EnableVpcClassicLinkResponse (Maybe Bool)

Returns true if the request succeeds; otherwise, it returns an error.

enableVpcClassicLinkResponse_httpStatus :: Lens' EnableVpcClassicLinkResponse Int

The response's http status code.