Network.AWS.EC2.ExportImage

Description

Exports an Amazon Machine Image (AMI) to a VM file. For more information, see Exporting a VM directly from an Amazon Machine Image (AMI) in the VM Import/Export User Guide.

Creating a Request

data ExportImage

See: newExportImage smart constructor.

Constructors

ExportImage'

Fields

tagSpecifications :: Maybe [TagSpecification]

The tags to apply to the export image task during creation.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

roleName :: Maybe Text

The name of the role that grants VM Import/Export permission to export images to your Amazon S3 bucket. If this parameter is not specified, the default role is named 'vmimport'.

description :: Maybe Text

A description of the image being exported. The maximum length is 255 characters.

clientToken :: Maybe Text

Token to enable idempotency for export image requests.

diskImageFormat :: DiskImageFormat

The disk image format.

imageId :: Text

The ID of the image.

s3ExportLocation :: ExportTaskS3LocationRequest

Information about the destination Amazon S3 bucket. The bucket must exist and grant WRITE and READ_ACP permissions to the Amazon Web Services account vm-import-export@amazon.com.

newExportImage

Arguments

:: DiskImageFormat
$sel:diskImageFormat:ExportImage'

-> Text
$sel:imageId:ExportImage'

-> ExportTaskS3LocationRequest
$sel:s3ExportLocation:ExportImage'

-> ExportImage

Create a value of ExportImage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tagSpecifications:ExportImage', exportImage_tagSpecifications - The tags to apply to the export image task during creation.

$sel:dryRun:ExportImage', exportImage_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:roleName:ExportImage', exportImage_roleName - The name of the role that grants VM Import/Export permission to export images to your Amazon S3 bucket. If this parameter is not specified, the default role is named 'vmimport'.

$sel:description:ExportImage', exportImage_description - A description of the image being exported. The maximum length is 255 characters.

$sel:clientToken:ExportImage', exportImage_clientToken - Token to enable idempotency for export image requests.

$sel:diskImageFormat:ExportImage', exportImage_diskImageFormat - The disk image format.

$sel:imageId:ExportImage', exportImage_imageId - The ID of the image.

$sel:s3ExportLocation:ExportImage', exportImage_s3ExportLocation - Information about the destination Amazon S3 bucket. The bucket must exist and grant WRITE and READ_ACP permissions to the Amazon Web Services account vm-import-export@amazon.com.

Request Lenses

exportImage_tagSpecifications :: Lens' ExportImage (Maybe [TagSpecification])

The tags to apply to the export image task during creation.

exportImage_dryRun :: Lens' ExportImage (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

exportImage_roleName :: Lens' ExportImage (Maybe Text)

The name of the role that grants VM Import/Export permission to export images to your Amazon S3 bucket. If this parameter is not specified, the default role is named 'vmimport'.

exportImage_description :: Lens' ExportImage (Maybe Text)

A description of the image being exported. The maximum length is 255 characters.

exportImage_clientToken :: Lens' ExportImage (Maybe Text)

Token to enable idempotency for export image requests.

exportImage_diskImageFormat :: Lens' ExportImage DiskImageFormat

The disk image format.

exportImage_imageId :: Lens' ExportImage Text

The ID of the image.

exportImage_s3ExportLocation :: Lens' ExportImage ExportTaskS3LocationRequest

Information about the destination Amazon S3 bucket. The bucket must exist and grant WRITE and READ_ACP permissions to the Amazon Web Services account vm-import-export@amazon.com.

Destructuring the Response

data ExportImageResponse

See: newExportImageResponse smart constructor.

Constructors

ExportImageResponse'

Fields

statusMessage :: Maybe Text

The status message for the export image task.

status :: Maybe Text

The status of the export image task. The possible values are active, completed, deleting, and deleted.

diskImageFormat :: Maybe DiskImageFormat

The disk image format for the exported image.

roleName :: Maybe Text

The name of the role that grants VM Import/Export permission to export images to your Amazon S3 bucket.

imageId :: Maybe Text

The ID of the image.

tags :: Maybe [Tag]

Any tags assigned to the export image task.

s3ExportLocation :: Maybe ExportTaskS3Location

Information about the destination Amazon S3 bucket.

description :: Maybe Text

A description of the image being exported.

exportImageTaskId :: Maybe Text

The ID of the export image task.

progress :: Maybe Text

The percent complete of the export image task.

httpStatus :: Int

The response's http status code.

newExportImageResponse

Arguments

:: Int
$sel:httpStatus:ExportImageResponse'

-> ExportImageResponse

Create a value of ExportImageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:statusMessage:ExportImageResponse', exportImageResponse_statusMessage - The status message for the export image task.

$sel:status:ExportImageResponse', exportImageResponse_status - The status of the export image task. The possible values are active, completed, deleting, and deleted.

$sel:diskImageFormat:ExportImage', exportImageResponse_diskImageFormat - The disk image format for the exported image.

$sel:roleName:ExportImage', exportImageResponse_roleName - The name of the role that grants VM Import/Export permission to export images to your Amazon S3 bucket.

$sel:imageId:ExportImage', exportImageResponse_imageId - The ID of the image.

$sel:tags:ExportImageResponse', exportImageResponse_tags - Any tags assigned to the export image task.

$sel:s3ExportLocation:ExportImage', exportImageResponse_s3ExportLocation - Information about the destination Amazon S3 bucket.

$sel:description:ExportImage', exportImageResponse_description - A description of the image being exported.

$sel:exportImageTaskId:ExportImageResponse', exportImageResponse_exportImageTaskId - The ID of the export image task.

$sel:progress:ExportImageResponse', exportImageResponse_progress - The percent complete of the export image task.

$sel:httpStatus:ExportImageResponse', exportImageResponse_httpStatus - The response's http status code.

Response Lenses

exportImageResponse_statusMessage :: Lens' ExportImageResponse (Maybe Text)

The status message for the export image task.

exportImageResponse_status :: Lens' ExportImageResponse (Maybe Text)

The status of the export image task. The possible values are active, completed, deleting, and deleted.

exportImageResponse_diskImageFormat :: Lens' ExportImageResponse (Maybe DiskImageFormat)

The disk image format for the exported image.

exportImageResponse_roleName :: Lens' ExportImageResponse (Maybe Text)

The name of the role that grants VM Import/Export permission to export images to your Amazon S3 bucket.

exportImageResponse_imageId :: Lens' ExportImageResponse (Maybe Text)

The ID of the image.

exportImageResponse_tags :: Lens' ExportImageResponse (Maybe [Tag])

Any tags assigned to the export image task.

exportImageResponse_s3ExportLocation :: Lens' ExportImageResponse (Maybe ExportTaskS3Location)

Information about the destination Amazon S3 bucket.

exportImageResponse_description :: Lens' ExportImageResponse (Maybe Text)

A description of the image being exported.

exportImageResponse_exportImageTaskId :: Lens' ExportImageResponse (Maybe Text)

The ID of the export image task.

exportImageResponse_progress :: Lens' ExportImageResponse (Maybe Text)

The percent complete of the export image task.

exportImageResponse_httpStatus :: Lens' ExportImageResponse Int

The response's http status code.