Network.AWS.EC2.GetCapacityReservationUsage

Description

Gets usage information about a Capacity Reservation. If the Capacity Reservation is shared, it shows usage information for the Capacity Reservation owner and each Amazon Web Services account that is currently using the shared capacity. If the Capacity Reservation is not shared, it shows only the Capacity Reservation owner's usage.

Creating a Request

data GetCapacityReservationUsage

See: newGetCapacityReservationUsage smart constructor.

Constructors

GetCapacityReservationUsage'

Fields

nextToken :: Maybe Text

The token to use to retrieve the next page of results.

maxResults :: Maybe Natural

The maximum number of results to return for the request in a single page. The remaining results can be seen by sending another request with the returned nextToken value. This value can be between 5 and 500. If maxResults is given a larger value than 500, you receive an error.

Valid range: Minimum value of 1. Maximum value of 1000.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

capacityReservationId :: Text

The ID of the Capacity Reservation.

newGetCapacityReservationUsage

Arguments

:: Text
$sel:capacityReservationId:GetCapacityReservationUsage'

-> GetCapacityReservationUsage

Create a value of GetCapacityReservationUsage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetCapacityReservationUsage', getCapacityReservationUsage_nextToken - The token to use to retrieve the next page of results.

$sel:maxResults:GetCapacityReservationUsage', getCapacityReservationUsage_maxResults - The maximum number of results to return for the request in a single page. The remaining results can be seen by sending another request with the returned nextToken value. This value can be between 5 and 500. If maxResults is given a larger value than 500, you receive an error.

Valid range: Minimum value of 1. Maximum value of 1000.

$sel:dryRun:GetCapacityReservationUsage', getCapacityReservationUsage_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:capacityReservationId:GetCapacityReservationUsage', getCapacityReservationUsage_capacityReservationId - The ID of the Capacity Reservation.

Request Lenses

getCapacityReservationUsage_nextToken :: Lens' GetCapacityReservationUsage (Maybe Text)

The token to use to retrieve the next page of results.

getCapacityReservationUsage_maxResults :: Lens' GetCapacityReservationUsage (Maybe Natural)

The maximum number of results to return for the request in a single page. The remaining results can be seen by sending another request with the returned nextToken value. This value can be between 5 and 500. If maxResults is given a larger value than 500, you receive an error.

Valid range: Minimum value of 1. Maximum value of 1000.

getCapacityReservationUsage_dryRun :: Lens' GetCapacityReservationUsage (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

getCapacityReservationUsage_capacityReservationId :: Lens' GetCapacityReservationUsage Text

The ID of the Capacity Reservation.

Destructuring the Response

data GetCapacityReservationUsageResponse

See: newGetCapacityReservationUsageResponse smart constructor.

Constructors

GetCapacityReservationUsageResponse'

Fields

instanceUsages :: Maybe [InstanceUsage]

Information about the Capacity Reservation usage.

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

instanceType :: Maybe Text

The type of instance for which the Capacity Reservation reserves capacity.

availableInstanceCount :: Maybe Int

The remaining capacity. Indicates the number of instances that can be launched in the Capacity Reservation.

state :: Maybe CapacityReservationState

The current state of the Capacity Reservation. A Capacity Reservation can be in one of the following states:

active - The Capacity Reservation is active and the capacity is available for your use.

expired - The Capacity Reservation expired automatically at the date and time specified in your request. The reserved capacity is no longer available for your use.

cancelled - The Capacity Reservation was cancelled. The reserved capacity is no longer available for your use.

pending - The Capacity Reservation request was successful but the capacity provisioning is still pending.

failed - The Capacity Reservation request has failed. A request might fail due to invalid request parameters, capacity constraints, or instance limit constraints. Failed requests are retained for 60 minutes.

capacityReservationId :: Maybe Text

The ID of the Capacity Reservation.

totalInstanceCount :: Maybe Int

The number of instances for which the Capacity Reservation reserves capacity.

httpStatus :: Int

The response's http status code.

newGetCapacityReservationUsageResponse

Arguments

:: Int
$sel:httpStatus:GetCapacityReservationUsageResponse'

-> GetCapacityReservationUsageResponse

Create a value of GetCapacityReservationUsageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceUsages:GetCapacityReservationUsageResponse', getCapacityReservationUsageResponse_instanceUsages - Information about the Capacity Reservation usage.

$sel:nextToken:GetCapacityReservationUsage', getCapacityReservationUsageResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:instanceType:GetCapacityReservationUsageResponse', getCapacityReservationUsageResponse_instanceType - The type of instance for which the Capacity Reservation reserves capacity.

$sel:availableInstanceCount:GetCapacityReservationUsageResponse', getCapacityReservationUsageResponse_availableInstanceCount - The remaining capacity. Indicates the number of instances that can be launched in the Capacity Reservation.

$sel:state:GetCapacityReservationUsageResponse', getCapacityReservationUsageResponse_state - The current state of the Capacity Reservation. A Capacity Reservation can be in one of the following states:

active - The Capacity Reservation is active and the capacity is available for your use.

expired - The Capacity Reservation expired automatically at the date and time specified in your request. The reserved capacity is no longer available for your use.

cancelled - The Capacity Reservation was cancelled. The reserved capacity is no longer available for your use.

pending - The Capacity Reservation request was successful but the capacity provisioning is still pending.

failed - The Capacity Reservation request has failed. A request might fail due to invalid request parameters, capacity constraints, or instance limit constraints. Failed requests are retained for 60 minutes.

$sel:capacityReservationId:GetCapacityReservationUsage', getCapacityReservationUsageResponse_capacityReservationId - The ID of the Capacity Reservation.

$sel:totalInstanceCount:GetCapacityReservationUsageResponse', getCapacityReservationUsageResponse_totalInstanceCount - The number of instances for which the Capacity Reservation reserves capacity.

$sel:httpStatus:GetCapacityReservationUsageResponse', getCapacityReservationUsageResponse_httpStatus - The response's http status code.

Response Lenses

getCapacityReservationUsageResponse_instanceUsages :: Lens' GetCapacityReservationUsageResponse (Maybe [InstanceUsage])

Information about the Capacity Reservation usage.

getCapacityReservationUsageResponse_nextToken :: Lens' GetCapacityReservationUsageResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

getCapacityReservationUsageResponse_instanceType :: Lens' GetCapacityReservationUsageResponse (Maybe Text)

The type of instance for which the Capacity Reservation reserves capacity.

getCapacityReservationUsageResponse_availableInstanceCount :: Lens' GetCapacityReservationUsageResponse (Maybe Int)

The remaining capacity. Indicates the number of instances that can be launched in the Capacity Reservation.

getCapacityReservationUsageResponse_state :: Lens' GetCapacityReservationUsageResponse (Maybe CapacityReservationState)

The current state of the Capacity Reservation. A Capacity Reservation can be in one of the following states:

active - The Capacity Reservation is active and the capacity is available for your use.

expired - The Capacity Reservation expired automatically at the date and time specified in your request. The reserved capacity is no longer available for your use.

cancelled - The Capacity Reservation was cancelled. The reserved capacity is no longer available for your use.

pending - The Capacity Reservation request was successful but the capacity provisioning is still pending.

failed - The Capacity Reservation request has failed. A request might fail due to invalid request parameters, capacity constraints, or instance limit constraints. Failed requests are retained for 60 minutes.

getCapacityReservationUsageResponse_capacityReservationId :: Lens' GetCapacityReservationUsageResponse (Maybe Text)

The ID of the Capacity Reservation.

getCapacityReservationUsageResponse_totalInstanceCount :: Lens' GetCapacityReservationUsageResponse (Maybe Int)

The number of instances for which the Capacity Reservation reserves capacity.

getCapacityReservationUsageResponse_httpStatus :: Lens' GetCapacityReservationUsageResponse Int

The response's http status code.