Network.AWS.EC2.GetConsoleOutput

Description

Gets the console output for the specified instance. For Linux instances, the instance console output displays the exact console output that would normally be displayed on a physical monitor attached to a computer. For Windows instances, the instance console output includes the last three system event log errors.

By default, the console output returns buffered information that was posted shortly after an instance transition state (start, stop, reboot, or terminate). This information is available for at least one hour after the most recent post. Only the most recent 64 KB of console output is available.

You can optionally retrieve the latest serial console output at any time during the instance lifecycle. This option is supported on instance types that use the Nitro hypervisor.

For more information, see Instance console output in the Amazon EC2 User Guide.

Creating a Request

data GetConsoleOutput

See: newGetConsoleOutput smart constructor.

Constructors

GetConsoleOutput'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

latest :: Maybe Bool

When enabled, retrieves the latest console output for the instance.

Default: disabled (false)

instanceId :: Text

The ID of the instance.

newGetConsoleOutput

Arguments

:: Text
$sel:instanceId:GetConsoleOutput'

-> GetConsoleOutput

Create a value of GetConsoleOutput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:GetConsoleOutput', getConsoleOutput_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:latest:GetConsoleOutput', getConsoleOutput_latest - When enabled, retrieves the latest console output for the instance.

Default: disabled (false)

$sel:instanceId:GetConsoleOutput', getConsoleOutput_instanceId - The ID of the instance.

Request Lenses

getConsoleOutput_dryRun :: Lens' GetConsoleOutput (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

getConsoleOutput_latest :: Lens' GetConsoleOutput (Maybe Bool)

When enabled, retrieves the latest console output for the instance.

Default: disabled (false)

getConsoleOutput_instanceId :: Lens' GetConsoleOutput Text

The ID of the instance.

Destructuring the Response

data GetConsoleOutputResponse

See: newGetConsoleOutputResponse smart constructor.

Constructors

GetConsoleOutputResponse'

Fields

instanceId :: Maybe Text

The ID of the instance.

output :: Maybe Text

The console output, base64-encoded. If you are using a command line tool, the tool decodes the output for you.

timestamp :: Maybe ISO8601

The time at which the output was last updated.

httpStatus :: Int

The response's http status code.

newGetConsoleOutputResponse

Arguments

:: Int
$sel:httpStatus:GetConsoleOutputResponse'

-> GetConsoleOutputResponse

Create a value of GetConsoleOutputResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceId:GetConsoleOutput', getConsoleOutputResponse_instanceId - The ID of the instance.

$sel:output:GetConsoleOutputResponse', getConsoleOutputResponse_output - The console output, base64-encoded. If you are using a command line tool, the tool decodes the output for you.

$sel:timestamp:GetConsoleOutputResponse', getConsoleOutputResponse_timestamp - The time at which the output was last updated.

$sel:httpStatus:GetConsoleOutputResponse', getConsoleOutputResponse_httpStatus - The response's http status code.

Response Lenses

getConsoleOutputResponse_instanceId :: Lens' GetConsoleOutputResponse (Maybe Text)

The ID of the instance.

getConsoleOutputResponse_output :: Lens' GetConsoleOutputResponse (Maybe Text)

The console output, base64-encoded. If you are using a command line tool, the tool decodes the output for you.

getConsoleOutputResponse_timestamp :: Lens' GetConsoleOutputResponse (Maybe UTCTime)

The time at which the output was last updated.

getConsoleOutputResponse_httpStatus :: Lens' GetConsoleOutputResponse Int

The response's http status code.