Network.AWS.EC2.GetConsoleScreenshot

Description

Retrieve a JPG-format screenshot of a running instance to help with troubleshooting.

The returned content is Base64-encoded.

Creating a Request

data GetConsoleScreenshot

See: newGetConsoleScreenshot smart constructor.

Constructors

GetConsoleScreenshot'

Fields

wakeUp :: Maybe Bool

When set to true, acts as keystroke input and wakes up an instance that's in standby or "sleep" mode.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

instanceId :: Text

The ID of the instance.

newGetConsoleScreenshot

Arguments

:: Text
$sel:instanceId:GetConsoleScreenshot'

-> GetConsoleScreenshot

Create a value of GetConsoleScreenshot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:wakeUp:GetConsoleScreenshot', getConsoleScreenshot_wakeUp - When set to true, acts as keystroke input and wakes up an instance that's in standby or "sleep" mode.

$sel:dryRun:GetConsoleScreenshot', getConsoleScreenshot_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:instanceId:GetConsoleScreenshot', getConsoleScreenshot_instanceId - The ID of the instance.

Request Lenses

getConsoleScreenshot_wakeUp :: Lens' GetConsoleScreenshot (Maybe Bool)

When set to true, acts as keystroke input and wakes up an instance that's in standby or "sleep" mode.

getConsoleScreenshot_dryRun :: Lens' GetConsoleScreenshot (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

getConsoleScreenshot_instanceId :: Lens' GetConsoleScreenshot Text

The ID of the instance.

Destructuring the Response

data GetConsoleScreenshotResponse

See: newGetConsoleScreenshotResponse smart constructor.

Constructors

GetConsoleScreenshotResponse'

Fields

instanceId :: Maybe Text

The ID of the instance.

imageData :: Maybe Text

The data that comprises the image.

httpStatus :: Int

The response's http status code.

newGetConsoleScreenshotResponse

Arguments

:: Int
$sel:httpStatus:GetConsoleScreenshotResponse'

-> GetConsoleScreenshotResponse

Create a value of GetConsoleScreenshotResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceId:GetConsoleScreenshot', getConsoleScreenshotResponse_instanceId - The ID of the instance.

$sel:imageData:GetConsoleScreenshotResponse', getConsoleScreenshotResponse_imageData - The data that comprises the image.

$sel:httpStatus:GetConsoleScreenshotResponse', getConsoleScreenshotResponse_httpStatus - The response's http status code.

Response Lenses

getConsoleScreenshotResponse_instanceId :: Lens' GetConsoleScreenshotResponse (Maybe Text)

The ID of the instance.

getConsoleScreenshotResponse_imageData :: Lens' GetConsoleScreenshotResponse (Maybe Text)

The data that comprises the image.

getConsoleScreenshotResponse_httpStatus :: Lens' GetConsoleScreenshotResponse Int

The response's http status code.