Network.AWS.EC2.GetEbsEncryptionByDefault

Description

Describes whether EBS encryption by default is enabled for your account in the current Region.

For more information, see Amazon EBS encryption in the Amazon Elastic Compute Cloud User Guide.

Creating a Request

data GetEbsEncryptionByDefault

See: newGetEbsEncryptionByDefault smart constructor.

Constructors

GetEbsEncryptionByDefault'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

newGetEbsEncryptionByDefault :: GetEbsEncryptionByDefault

Create a value of GetEbsEncryptionByDefault with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:GetEbsEncryptionByDefault', getEbsEncryptionByDefault_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

Request Lenses

getEbsEncryptionByDefault_dryRun :: Lens' GetEbsEncryptionByDefault (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

Destructuring the Response

data GetEbsEncryptionByDefaultResponse

See: newGetEbsEncryptionByDefaultResponse smart constructor.

Constructors

GetEbsEncryptionByDefaultResponse'

Fields

ebsEncryptionByDefault :: Maybe Bool

Indicates whether encryption by default is enabled.

httpStatus :: Int

The response's http status code.

newGetEbsEncryptionByDefaultResponse

Arguments

:: Int
$sel:httpStatus:GetEbsEncryptionByDefaultResponse'

-> GetEbsEncryptionByDefaultResponse

Create a value of GetEbsEncryptionByDefaultResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ebsEncryptionByDefault:GetEbsEncryptionByDefaultResponse', getEbsEncryptionByDefaultResponse_ebsEncryptionByDefault - Indicates whether encryption by default is enabled.

$sel:httpStatus:GetEbsEncryptionByDefaultResponse', getEbsEncryptionByDefaultResponse_httpStatus - The response's http status code.

Response Lenses

getEbsEncryptionByDefaultResponse_ebsEncryptionByDefault :: Lens' GetEbsEncryptionByDefaultResponse (Maybe Bool)

Indicates whether encryption by default is enabled.

getEbsEncryptionByDefaultResponse_httpStatus :: Lens' GetEbsEncryptionByDefaultResponse Int

The response's http status code.