Network.AWS.EC2.GetManagedPrefixListAssociations

Description

Gets information about the resources that are associated with the specified managed prefix list.

This operation returns paginated results.

Creating a Request

data GetManagedPrefixListAssociations

See: newGetManagedPrefixListAssociations smart constructor.

Constructors

GetManagedPrefixListAssociations'

Fields

nextToken :: Maybe Text

The token for the next page of results.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

maxResults :: Maybe Natural

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

prefixListId :: Text

The ID of the prefix list.

newGetManagedPrefixListAssociations

Arguments

:: Text
$sel:prefixListId:GetManagedPrefixListAssociations'

-> GetManagedPrefixListAssociations

Create a value of GetManagedPrefixListAssociations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetManagedPrefixListAssociations', getManagedPrefixListAssociations_nextToken - The token for the next page of results.

$sel:dryRun:GetManagedPrefixListAssociations', getManagedPrefixListAssociations_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:maxResults:GetManagedPrefixListAssociations', getManagedPrefixListAssociations_maxResults - The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

$sel:prefixListId:GetManagedPrefixListAssociations', getManagedPrefixListAssociations_prefixListId - The ID of the prefix list.

Request Lenses

getManagedPrefixListAssociations_nextToken :: Lens' GetManagedPrefixListAssociations (Maybe Text)

The token for the next page of results.

getManagedPrefixListAssociations_dryRun :: Lens' GetManagedPrefixListAssociations (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

getManagedPrefixListAssociations_maxResults :: Lens' GetManagedPrefixListAssociations (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

getManagedPrefixListAssociations_prefixListId :: Lens' GetManagedPrefixListAssociations Text

The ID of the prefix list.

Destructuring the Response

data GetManagedPrefixListAssociationsResponse

See: newGetManagedPrefixListAssociationsResponse smart constructor.

Constructors

GetManagedPrefixListAssociationsResponse'

Fields

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

prefixListAssociations :: Maybe [PrefixListAssociation]

Information about the associations.

httpStatus :: Int

The response's http status code.

newGetManagedPrefixListAssociationsResponse

Arguments

:: Int
$sel:httpStatus:GetManagedPrefixListAssociationsResponse'

-> GetManagedPrefixListAssociationsResponse

Create a value of GetManagedPrefixListAssociationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetManagedPrefixListAssociations', getManagedPrefixListAssociationsResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:prefixListAssociations:GetManagedPrefixListAssociationsResponse', getManagedPrefixListAssociationsResponse_prefixListAssociations - Information about the associations.

$sel:httpStatus:GetManagedPrefixListAssociationsResponse', getManagedPrefixListAssociationsResponse_httpStatus - The response's http status code.

Response Lenses

getManagedPrefixListAssociationsResponse_nextToken :: Lens' GetManagedPrefixListAssociationsResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

getManagedPrefixListAssociationsResponse_prefixListAssociations :: Lens' GetManagedPrefixListAssociationsResponse (Maybe [PrefixListAssociation])

Information about the associations.

getManagedPrefixListAssociationsResponse_httpStatus :: Lens' GetManagedPrefixListAssociationsResponse Int

The response's http status code.