Network.AWS.EC2.GetReservedInstancesExchangeQuote

Description

Returns a quote and exchange information for exchanging one or more specified Convertible Reserved Instances for a new Convertible Reserved Instance. If the exchange cannot be performed, the reason is returned in the response. Use AcceptReservedInstancesExchangeQuote to perform the exchange.

Creating a Request

data GetReservedInstancesExchangeQuote

Contains the parameters for GetReservedInstanceExchangeQuote.

See: newGetReservedInstancesExchangeQuote smart constructor.

Constructors

GetReservedInstancesExchangeQuote'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

targetConfigurations :: Maybe [TargetConfigurationRequest]

The configuration of the target Convertible Reserved Instance to exchange for your current Convertible Reserved Instances.

reservedInstanceIds :: [Text]

The IDs of the Convertible Reserved Instances to exchange.

newGetReservedInstancesExchangeQuote :: GetReservedInstancesExchangeQuote

Create a value of GetReservedInstancesExchangeQuote with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:GetReservedInstancesExchangeQuote', getReservedInstancesExchangeQuote_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:targetConfigurations:GetReservedInstancesExchangeQuote', getReservedInstancesExchangeQuote_targetConfigurations - The configuration of the target Convertible Reserved Instance to exchange for your current Convertible Reserved Instances.

$sel:reservedInstanceIds:GetReservedInstancesExchangeQuote', getReservedInstancesExchangeQuote_reservedInstanceIds - The IDs of the Convertible Reserved Instances to exchange.

Request Lenses

getReservedInstancesExchangeQuote_dryRun :: Lens' GetReservedInstancesExchangeQuote (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

getReservedInstancesExchangeQuote_targetConfigurations :: Lens' GetReservedInstancesExchangeQuote (Maybe [TargetConfigurationRequest])

The configuration of the target Convertible Reserved Instance to exchange for your current Convertible Reserved Instances.

getReservedInstancesExchangeQuote_reservedInstanceIds :: Lens' GetReservedInstancesExchangeQuote [Text]

The IDs of the Convertible Reserved Instances to exchange.

Destructuring the Response

data GetReservedInstancesExchangeQuoteResponse

Contains the output of GetReservedInstancesExchangeQuote.

See: newGetReservedInstancesExchangeQuoteResponse smart constructor.

Constructors

GetReservedInstancesExchangeQuoteResponse'

Fields

isValidExchange :: Maybe Bool

If true, the exchange is valid. If false, the exchange cannot be completed.

paymentDue :: Maybe Text

The total true upfront charge for the exchange.

reservedInstanceValueRollup :: Maybe ReservationValue

The cost associated with the Reserved Instance.

currencyCode :: Maybe Text

The currency of the transaction.

targetConfigurationValueRollup :: Maybe ReservationValue

The cost associated with the Reserved Instance.

validationFailureReason :: Maybe Text

Describes the reason why the exchange cannot be completed.

reservedInstanceValueSet :: Maybe [ReservedInstanceReservationValue]

The configuration of your Convertible Reserved Instances.

outputReservedInstancesWillExpireAt :: Maybe ISO8601

The new end date of the reservation term.

targetConfigurationValueSet :: Maybe [TargetReservationValue]

The values of the target Convertible Reserved Instances.

httpStatus :: Int

The response's http status code.

newGetReservedInstancesExchangeQuoteResponse

Arguments

:: Int
$sel:httpStatus:GetReservedInstancesExchangeQuoteResponse'

-> GetReservedInstancesExchangeQuoteResponse

Create a value of GetReservedInstancesExchangeQuoteResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:isValidExchange:GetReservedInstancesExchangeQuoteResponse', getReservedInstancesExchangeQuoteResponse_isValidExchange - If true, the exchange is valid. If false, the exchange cannot be completed.

$sel:paymentDue:GetReservedInstancesExchangeQuoteResponse', getReservedInstancesExchangeQuoteResponse_paymentDue - The total true upfront charge for the exchange.

$sel:reservedInstanceValueRollup:GetReservedInstancesExchangeQuoteResponse', getReservedInstancesExchangeQuoteResponse_reservedInstanceValueRollup - The cost associated with the Reserved Instance.

$sel:currencyCode:GetReservedInstancesExchangeQuoteResponse', getReservedInstancesExchangeQuoteResponse_currencyCode - The currency of the transaction.

$sel:targetConfigurationValueRollup:GetReservedInstancesExchangeQuoteResponse', getReservedInstancesExchangeQuoteResponse_targetConfigurationValueRollup - The cost associated with the Reserved Instance.

$sel:validationFailureReason:GetReservedInstancesExchangeQuoteResponse', getReservedInstancesExchangeQuoteResponse_validationFailureReason - Describes the reason why the exchange cannot be completed.

$sel:reservedInstanceValueSet:GetReservedInstancesExchangeQuoteResponse', getReservedInstancesExchangeQuoteResponse_reservedInstanceValueSet - The configuration of your Convertible Reserved Instances.

$sel:outputReservedInstancesWillExpireAt:GetReservedInstancesExchangeQuoteResponse', getReservedInstancesExchangeQuoteResponse_outputReservedInstancesWillExpireAt - The new end date of the reservation term.

$sel:targetConfigurationValueSet:GetReservedInstancesExchangeQuoteResponse', getReservedInstancesExchangeQuoteResponse_targetConfigurationValueSet - The values of the target Convertible Reserved Instances.

$sel:httpStatus:GetReservedInstancesExchangeQuoteResponse', getReservedInstancesExchangeQuoteResponse_httpStatus - The response's http status code.

Response Lenses

getReservedInstancesExchangeQuoteResponse_isValidExchange :: Lens' GetReservedInstancesExchangeQuoteResponse (Maybe Bool)

If true, the exchange is valid. If false, the exchange cannot be completed.

getReservedInstancesExchangeQuoteResponse_paymentDue :: Lens' GetReservedInstancesExchangeQuoteResponse (Maybe Text)

The total true upfront charge for the exchange.

getReservedInstancesExchangeQuoteResponse_reservedInstanceValueRollup :: Lens' GetReservedInstancesExchangeQuoteResponse (Maybe ReservationValue)

The cost associated with the Reserved Instance.

getReservedInstancesExchangeQuoteResponse_currencyCode :: Lens' GetReservedInstancesExchangeQuoteResponse (Maybe Text)

The currency of the transaction.

getReservedInstancesExchangeQuoteResponse_targetConfigurationValueRollup :: Lens' GetReservedInstancesExchangeQuoteResponse (Maybe ReservationValue)

The cost associated with the Reserved Instance.

getReservedInstancesExchangeQuoteResponse_validationFailureReason :: Lens' GetReservedInstancesExchangeQuoteResponse (Maybe Text)

Describes the reason why the exchange cannot be completed.

getReservedInstancesExchangeQuoteResponse_reservedInstanceValueSet :: Lens' GetReservedInstancesExchangeQuoteResponse (Maybe [ReservedInstanceReservationValue])

The configuration of your Convertible Reserved Instances.

getReservedInstancesExchangeQuoteResponse_outputReservedInstancesWillExpireAt :: Lens' GetReservedInstancesExchangeQuoteResponse (Maybe UTCTime)

The new end date of the reservation term.

getReservedInstancesExchangeQuoteResponse_targetConfigurationValueSet :: Lens' GetReservedInstancesExchangeQuoteResponse (Maybe [TargetReservationValue])

The values of the target Convertible Reserved Instances.

getReservedInstancesExchangeQuoteResponse_httpStatus :: Lens' GetReservedInstancesExchangeQuoteResponse Int

The response's http status code.