Network.AWS.EC2.GetSerialConsoleAccessStatus

Description

Retrieves the access status of your account to the EC2 serial console of all instances. By default, access to the EC2 serial console is disabled for your account. For more information, see Manage account access to the EC2 serial console in the Amazon EC2 User Guide.

Creating a Request

data GetSerialConsoleAccessStatus

See: newGetSerialConsoleAccessStatus smart constructor.

Constructors

GetSerialConsoleAccessStatus'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

newGetSerialConsoleAccessStatus :: GetSerialConsoleAccessStatus

Create a value of GetSerialConsoleAccessStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:GetSerialConsoleAccessStatus', getSerialConsoleAccessStatus_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

Request Lenses

getSerialConsoleAccessStatus_dryRun :: Lens' GetSerialConsoleAccessStatus (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

Destructuring the Response

data GetSerialConsoleAccessStatusResponse

See: newGetSerialConsoleAccessStatusResponse smart constructor.

Constructors

GetSerialConsoleAccessStatusResponse'

Fields

serialConsoleAccessEnabled :: Maybe Bool

If true, access to the EC2 serial console of all instances is enabled for your account. If false, access to the EC2 serial console of all instances is disabled for your account.

httpStatus :: Int

The response's http status code.

newGetSerialConsoleAccessStatusResponse

Arguments

:: Int
$sel:httpStatus:GetSerialConsoleAccessStatusResponse'

-> GetSerialConsoleAccessStatusResponse

Create a value of GetSerialConsoleAccessStatusResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serialConsoleAccessEnabled:GetSerialConsoleAccessStatusResponse', getSerialConsoleAccessStatusResponse_serialConsoleAccessEnabled - If true, access to the EC2 serial console of all instances is enabled for your account. If false, access to the EC2 serial console of all instances is disabled for your account.

$sel:httpStatus:GetSerialConsoleAccessStatusResponse', getSerialConsoleAccessStatusResponse_httpStatus - The response's http status code.

Response Lenses

getSerialConsoleAccessStatusResponse_serialConsoleAccessEnabled :: Lens' GetSerialConsoleAccessStatusResponse (Maybe Bool)

If true, access to the EC2 serial console of all instances is enabled for your account. If false, access to the EC2 serial console of all instances is disabled for your account.

getSerialConsoleAccessStatusResponse_httpStatus :: Lens' GetSerialConsoleAccessStatusResponse Int

The response's http status code.