Network.AWS.EC2.GetSubnetCidrReservations

Description

Gets information about the subnet CIDR reservations.

Creating a Request

data GetSubnetCidrReservations

See: newGetSubnetCidrReservations smart constructor.

Constructors

GetSubnetCidrReservations'

Fields

nextToken :: Maybe Text

The token for the next page of results.

maxResults :: Maybe Natural

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

filters :: Maybe [Filter]

One or more filters.

reservationType - The type of reservation (prefix | explicit).

subnet-id - The ID of the subnet.

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

subnetId :: Text

The ID of the subnet.

newGetSubnetCidrReservations

Arguments

:: Text
$sel:subnetId:GetSubnetCidrReservations'

-> GetSubnetCidrReservations

Create a value of GetSubnetCidrReservations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetSubnetCidrReservations', getSubnetCidrReservations_nextToken - The token for the next page of results.

$sel:maxResults:GetSubnetCidrReservations', getSubnetCidrReservations_maxResults - The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

$sel:dryRun:GetSubnetCidrReservations', getSubnetCidrReservations_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:filters:GetSubnetCidrReservations', getSubnetCidrReservations_filters - One or more filters.

reservationType - The type of reservation (prefix | explicit).

subnet-id - The ID of the subnet.

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

$sel:subnetId:GetSubnetCidrReservations', getSubnetCidrReservations_subnetId - The ID of the subnet.

Request Lenses

getSubnetCidrReservations_nextToken :: Lens' GetSubnetCidrReservations (Maybe Text)

The token for the next page of results.

getSubnetCidrReservations_maxResults :: Lens' GetSubnetCidrReservations (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

getSubnetCidrReservations_dryRun :: Lens' GetSubnetCidrReservations (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

getSubnetCidrReservations_filters :: Lens' GetSubnetCidrReservations (Maybe [Filter])

One or more filters.

reservationType - The type of reservation (prefix | explicit).

subnet-id - The ID of the subnet.

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

getSubnetCidrReservations_subnetId :: Lens' GetSubnetCidrReservations Text

The ID of the subnet.

Destructuring the Response

data GetSubnetCidrReservationsResponse

See: newGetSubnetCidrReservationsResponse smart constructor.

Constructors

GetSubnetCidrReservationsResponse'

Fields

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

subnetIpv4CidrReservations :: Maybe [SubnetCidrReservation]

Information about the IPv4 subnet CIDR reservations.

subnetIpv6CidrReservations :: Maybe [SubnetCidrReservation]

Information about the IPv6 subnet CIDR reservations.

httpStatus :: Int

The response's http status code.

newGetSubnetCidrReservationsResponse

Arguments

:: Int
$sel:httpStatus:GetSubnetCidrReservationsResponse'

-> GetSubnetCidrReservationsResponse

Create a value of GetSubnetCidrReservationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetSubnetCidrReservations', getSubnetCidrReservationsResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:subnetIpv4CidrReservations:GetSubnetCidrReservationsResponse', getSubnetCidrReservationsResponse_subnetIpv4CidrReservations - Information about the IPv4 subnet CIDR reservations.

$sel:subnetIpv6CidrReservations:GetSubnetCidrReservationsResponse', getSubnetCidrReservationsResponse_subnetIpv6CidrReservations - Information about the IPv6 subnet CIDR reservations.

$sel:httpStatus:GetSubnetCidrReservationsResponse', getSubnetCidrReservationsResponse_httpStatus - The response's http status code.

Response Lenses

getSubnetCidrReservationsResponse_nextToken :: Lens' GetSubnetCidrReservationsResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

getSubnetCidrReservationsResponse_subnetIpv4CidrReservations :: Lens' GetSubnetCidrReservationsResponse (Maybe [SubnetCidrReservation])

Information about the IPv4 subnet CIDR reservations.

getSubnetCidrReservationsResponse_subnetIpv6CidrReservations :: Lens' GetSubnetCidrReservationsResponse (Maybe [SubnetCidrReservation])

Information about the IPv6 subnet CIDR reservations.

getSubnetCidrReservationsResponse_httpStatus :: Lens' GetSubnetCidrReservationsResponse Int

The response's http status code.