Network.AWS.EC2.GetTransitGatewayAttachmentPropagations

Description

Lists the route tables to which the specified resource attachment propagates routes.

This operation returns paginated results.

Creating a Request

data GetTransitGatewayAttachmentPropagations

See: newGetTransitGatewayAttachmentPropagations smart constructor.

Constructors

GetTransitGatewayAttachmentPropagations'

Fields

filters :: Maybe [Filter]

One or more filters. The possible values are:

transit-gateway-route-table-id - The ID of the transit gateway route table.

nextToken :: Maybe Text

The token for the next page of results.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

maxResults :: Maybe Natural

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

transitGatewayAttachmentId :: Text

The ID of the attachment.

newGetTransitGatewayAttachmentPropagations

Arguments

:: Text
$sel:transitGatewayAttachmentId:GetTransitGatewayAttachmentPropagations'

-> GetTransitGatewayAttachmentPropagations

Create a value of GetTransitGatewayAttachmentPropagations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:GetTransitGatewayAttachmentPropagations', getTransitGatewayAttachmentPropagations_filters - One or more filters. The possible values are:

transit-gateway-route-table-id - The ID of the transit gateway route table.

$sel:nextToken:GetTransitGatewayAttachmentPropagations', getTransitGatewayAttachmentPropagations_nextToken - The token for the next page of results.

$sel:dryRun:GetTransitGatewayAttachmentPropagations', getTransitGatewayAttachmentPropagations_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:maxResults:GetTransitGatewayAttachmentPropagations', getTransitGatewayAttachmentPropagations_maxResults - The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

$sel:transitGatewayAttachmentId:GetTransitGatewayAttachmentPropagations', getTransitGatewayAttachmentPropagations_transitGatewayAttachmentId - The ID of the attachment.

Request Lenses

getTransitGatewayAttachmentPropagations_filters :: Lens' GetTransitGatewayAttachmentPropagations (Maybe [Filter])

One or more filters. The possible values are:

transit-gateway-route-table-id - The ID of the transit gateway route table.

getTransitGatewayAttachmentPropagations_nextToken :: Lens' GetTransitGatewayAttachmentPropagations (Maybe Text)

The token for the next page of results.

getTransitGatewayAttachmentPropagations_dryRun :: Lens' GetTransitGatewayAttachmentPropagations (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

getTransitGatewayAttachmentPropagations_maxResults :: Lens' GetTransitGatewayAttachmentPropagations (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

getTransitGatewayAttachmentPropagations_transitGatewayAttachmentId :: Lens' GetTransitGatewayAttachmentPropagations Text

The ID of the attachment.

Destructuring the Response

data GetTransitGatewayAttachmentPropagationsResponse

See: newGetTransitGatewayAttachmentPropagationsResponse smart constructor.

Constructors

GetTransitGatewayAttachmentPropagationsResponse'

Fields

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

transitGatewayAttachmentPropagations :: Maybe [TransitGatewayAttachmentPropagation]

Information about the propagation route tables.

httpStatus :: Int

The response's http status code.

newGetTransitGatewayAttachmentPropagationsResponse

Arguments

:: Int
$sel:httpStatus:GetTransitGatewayAttachmentPropagationsResponse'

-> GetTransitGatewayAttachmentPropagationsResponse

Create a value of GetTransitGatewayAttachmentPropagationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetTransitGatewayAttachmentPropagations', getTransitGatewayAttachmentPropagationsResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:transitGatewayAttachmentPropagations:GetTransitGatewayAttachmentPropagationsResponse', getTransitGatewayAttachmentPropagationsResponse_transitGatewayAttachmentPropagations - Information about the propagation route tables.

$sel:httpStatus:GetTransitGatewayAttachmentPropagationsResponse', getTransitGatewayAttachmentPropagationsResponse_httpStatus - The response's http status code.

Response Lenses

getTransitGatewayAttachmentPropagationsResponse_nextToken :: Lens' GetTransitGatewayAttachmentPropagationsResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

getTransitGatewayAttachmentPropagationsResponse_transitGatewayAttachmentPropagations :: Lens' GetTransitGatewayAttachmentPropagationsResponse (Maybe [TransitGatewayAttachmentPropagation])

Information about the propagation route tables.

getTransitGatewayAttachmentPropagationsResponse_httpStatus :: Lens' GetTransitGatewayAttachmentPropagationsResponse Int

The response's http status code.