Network.AWS.EC2.GetTransitGatewayMulticastDomainAssociations

Description

Gets information about the associations for the transit gateway multicast domain.

This operation returns paginated results.

Creating a Request

data GetTransitGatewayMulticastDomainAssociations

See: newGetTransitGatewayMulticastDomainAssociations smart constructor.

Constructors

GetTransitGatewayMulticastDomainAssociations'

Fields

nextToken :: Maybe Text

The token for the next page of results.

maxResults :: Maybe Natural

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

transitGatewayMulticastDomainId :: Maybe Text

The ID of the transit gateway multicast domain.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

filters :: Maybe [Filter]

One or more filters. The possible values are:

resource-id - The ID of the resource.

resource-type - The type of resource. The valid value is: vpc.

state - The state of the subnet association. Valid values are associated | associating | disassociated | disassociating.

subnet-id - The ID of the subnet.

transit-gateway-attachment-id - The id of the transit gateway attachment.

newGetTransitGatewayMulticastDomainAssociations :: GetTransitGatewayMulticastDomainAssociations

Create a value of GetTransitGatewayMulticastDomainAssociations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetTransitGatewayMulticastDomainAssociations', getTransitGatewayMulticastDomainAssociations_nextToken - The token for the next page of results.

$sel:maxResults:GetTransitGatewayMulticastDomainAssociations', getTransitGatewayMulticastDomainAssociations_maxResults - The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

$sel:transitGatewayMulticastDomainId:GetTransitGatewayMulticastDomainAssociations', getTransitGatewayMulticastDomainAssociations_transitGatewayMulticastDomainId - The ID of the transit gateway multicast domain.

$sel:dryRun:GetTransitGatewayMulticastDomainAssociations', getTransitGatewayMulticastDomainAssociations_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:filters:GetTransitGatewayMulticastDomainAssociations', getTransitGatewayMulticastDomainAssociations_filters - One or more filters. The possible values are:

resource-id - The ID of the resource.

resource-type - The type of resource. The valid value is: vpc.

state - The state of the subnet association. Valid values are associated | associating | disassociated | disassociating.

subnet-id - The ID of the subnet.

transit-gateway-attachment-id - The id of the transit gateway attachment.

Request Lenses

getTransitGatewayMulticastDomainAssociations_nextToken :: Lens' GetTransitGatewayMulticastDomainAssociations (Maybe Text)

The token for the next page of results.

getTransitGatewayMulticastDomainAssociations_maxResults :: Lens' GetTransitGatewayMulticastDomainAssociations (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

getTransitGatewayMulticastDomainAssociations_transitGatewayMulticastDomainId :: Lens' GetTransitGatewayMulticastDomainAssociations (Maybe Text)

The ID of the transit gateway multicast domain.

getTransitGatewayMulticastDomainAssociations_dryRun :: Lens' GetTransitGatewayMulticastDomainAssociations (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

getTransitGatewayMulticastDomainAssociations_filters :: Lens' GetTransitGatewayMulticastDomainAssociations (Maybe [Filter])

One or more filters. The possible values are:

resource-id - The ID of the resource.

resource-type - The type of resource. The valid value is: vpc.

state - The state of the subnet association. Valid values are associated | associating | disassociated | disassociating.

subnet-id - The ID of the subnet.

transit-gateway-attachment-id - The id of the transit gateway attachment.

Destructuring the Response

data GetTransitGatewayMulticastDomainAssociationsResponse

See: newGetTransitGatewayMulticastDomainAssociationsResponse smart constructor.

Constructors

GetTransitGatewayMulticastDomainAssociationsResponse'

Fields

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

multicastDomainAssociations :: Maybe [TransitGatewayMulticastDomainAssociation]

Information about the multicast domain associations.

httpStatus :: Int

The response's http status code.

newGetTransitGatewayMulticastDomainAssociationsResponse

Arguments

:: Int
$sel:httpStatus:GetTransitGatewayMulticastDomainAssociationsResponse'

-> GetTransitGatewayMulticastDomainAssociationsResponse

Create a value of GetTransitGatewayMulticastDomainAssociationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetTransitGatewayMulticastDomainAssociations', getTransitGatewayMulticastDomainAssociationsResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:multicastDomainAssociations:GetTransitGatewayMulticastDomainAssociationsResponse', getTransitGatewayMulticastDomainAssociationsResponse_multicastDomainAssociations - Information about the multicast domain associations.

$sel:httpStatus:GetTransitGatewayMulticastDomainAssociationsResponse', getTransitGatewayMulticastDomainAssociationsResponse_httpStatus - The response's http status code.

Response Lenses

getTransitGatewayMulticastDomainAssociationsResponse_nextToken :: Lens' GetTransitGatewayMulticastDomainAssociationsResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

getTransitGatewayMulticastDomainAssociationsResponse_multicastDomainAssociations :: Lens' GetTransitGatewayMulticastDomainAssociationsResponse (Maybe [TransitGatewayMulticastDomainAssociation])

Information about the multicast domain associations.

getTransitGatewayMulticastDomainAssociationsResponse_httpStatus :: Lens' GetTransitGatewayMulticastDomainAssociationsResponse Int

The response's http status code.