Network.AWS.EC2.ImportKeyPair

Description

Imports the public key from an RSA or ED25519 key pair that you created with a third-party tool. Compare this with CreateKeyPair, in which Amazon Web Services creates the key pair and gives the keys to you (Amazon Web Services keeps a copy of the public key). With ImportKeyPair, you create the key pair and give Amazon Web Services just the public key. The private key is never transferred between you and Amazon Web Services.

For more information about key pairs, see Amazon EC2 key pairs in the Amazon Elastic Compute Cloud User Guide.

Creating a Request

data ImportKeyPair

See: newImportKeyPair smart constructor.

Constructors

ImportKeyPair'

Fields

tagSpecifications :: Maybe [TagSpecification]

The tags to apply to the imported key pair.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

keyName :: Text

A unique name for the key pair.

publicKeyMaterial :: Base64

The public key. For API calls, the text must be base64-encoded. For command line tools, base64 encoding is performed for you.

newImportKeyPair

Arguments

:: Text
$sel:keyName:ImportKeyPair'

-> ByteString
$sel:publicKeyMaterial:ImportKeyPair'

-> ImportKeyPair

Create a value of ImportKeyPair with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tagSpecifications:ImportKeyPair', importKeyPair_tagSpecifications - The tags to apply to the imported key pair.

$sel:dryRun:ImportKeyPair', importKeyPair_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:keyName:ImportKeyPair', importKeyPair_keyName - A unique name for the key pair.

$sel:publicKeyMaterial:ImportKeyPair', importKeyPair_publicKeyMaterial - The public key. For API calls, the text must be base64-encoded. For command line tools, base64 encoding is performed for you.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

Request Lenses

importKeyPair_tagSpecifications :: Lens' ImportKeyPair (Maybe [TagSpecification])

The tags to apply to the imported key pair.

importKeyPair_dryRun :: Lens' ImportKeyPair (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

importKeyPair_keyName :: Lens' ImportKeyPair Text

A unique name for the key pair.

importKeyPair_publicKeyMaterial :: Lens' ImportKeyPair ByteString

The public key. For API calls, the text must be base64-encoded. For command line tools, base64 encoding is performed for you.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

Destructuring the Response

data ImportKeyPairResponse

See: newImportKeyPairResponse smart constructor.

Constructors

ImportKeyPairResponse'

Fields

keyFingerprint :: Maybe Text

The MD5 public key fingerprint as specified in section 4 of RFC 4716.

keyName :: Maybe Text

The key pair name that you provided.

keyPairId :: Maybe Text

The ID of the resulting key pair.

tags :: Maybe [Tag]

The tags applied to the imported key pair.

httpStatus :: Int

The response's http status code.

newImportKeyPairResponse

Arguments

:: Int
$sel:httpStatus:ImportKeyPairResponse'

-> ImportKeyPairResponse

Create a value of ImportKeyPairResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:keyFingerprint:ImportKeyPairResponse', importKeyPairResponse_keyFingerprint - The MD5 public key fingerprint as specified in section 4 of RFC 4716.

$sel:keyName:ImportKeyPair', importKeyPairResponse_keyName - The key pair name that you provided.

$sel:keyPairId:ImportKeyPairResponse', importKeyPairResponse_keyPairId - The ID of the resulting key pair.

$sel:tags:ImportKeyPairResponse', importKeyPairResponse_tags - The tags applied to the imported key pair.

$sel:httpStatus:ImportKeyPairResponse', importKeyPairResponse_httpStatus - The response's http status code.

Response Lenses

importKeyPairResponse_keyFingerprint :: Lens' ImportKeyPairResponse (Maybe Text)

The MD5 public key fingerprint as specified in section 4 of RFC 4716.

importKeyPairResponse_keyName :: Lens' ImportKeyPairResponse (Maybe Text)

The key pair name that you provided.

importKeyPairResponse_keyPairId :: Lens' ImportKeyPairResponse (Maybe Text)

The ID of the resulting key pair.

importKeyPairResponse_tags :: Lens' ImportKeyPairResponse (Maybe [Tag])

The tags applied to the imported key pair.

importKeyPairResponse_httpStatus :: Lens' ImportKeyPairResponse Int

The response's http status code.