Network.AWS.EC2.ModifyAvailabilityZoneGroup

Description

Changes the opt-in status of the Local Zone and Wavelength Zone group for your account.

Use DescribeAvailabilityZones to view the value for GroupName.

Creating a Request

data ModifyAvailabilityZoneGroup

See: newModifyAvailabilityZoneGroup smart constructor.

Constructors

ModifyAvailabilityZoneGroup'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

groupName :: Text

The name of the Availability Zone group, Local Zone group, or Wavelength Zone group.

optInStatus :: ModifyAvailabilityZoneOptInStatus

Indicates whether you are opted in to the Local Zone group or Wavelength Zone group. The only valid value is opted-in. You must contact AWS Support to opt out of a Local Zone group, or Wavelength Zone group.

newModifyAvailabilityZoneGroup

Arguments

:: Text
$sel:groupName:ModifyAvailabilityZoneGroup'

-> ModifyAvailabilityZoneOptInStatus
$sel:optInStatus:ModifyAvailabilityZoneGroup'

-> ModifyAvailabilityZoneGroup

Create a value of ModifyAvailabilityZoneGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:ModifyAvailabilityZoneGroup', modifyAvailabilityZoneGroup_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:groupName:ModifyAvailabilityZoneGroup', modifyAvailabilityZoneGroup_groupName - The name of the Availability Zone group, Local Zone group, or Wavelength Zone group.

$sel:optInStatus:ModifyAvailabilityZoneGroup', modifyAvailabilityZoneGroup_optInStatus - Indicates whether you are opted in to the Local Zone group or Wavelength Zone group. The only valid value is opted-in. You must contact AWS Support to opt out of a Local Zone group, or Wavelength Zone group.

Request Lenses

modifyAvailabilityZoneGroup_dryRun :: Lens' ModifyAvailabilityZoneGroup (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

modifyAvailabilityZoneGroup_groupName :: Lens' ModifyAvailabilityZoneGroup Text

The name of the Availability Zone group, Local Zone group, or Wavelength Zone group.

modifyAvailabilityZoneGroup_optInStatus :: Lens' ModifyAvailabilityZoneGroup ModifyAvailabilityZoneOptInStatus

Indicates whether you are opted in to the Local Zone group or Wavelength Zone group. The only valid value is opted-in. You must contact AWS Support to opt out of a Local Zone group, or Wavelength Zone group.

Destructuring the Response

data ModifyAvailabilityZoneGroupResponse

See: newModifyAvailabilityZoneGroupResponse smart constructor.

Constructors

ModifyAvailabilityZoneGroupResponse'

Fields

return' :: Maybe Bool

Is true if the request succeeds, and an error otherwise.

httpStatus :: Int

The response's http status code.

newModifyAvailabilityZoneGroupResponse

Arguments

:: Int
$sel:httpStatus:ModifyAvailabilityZoneGroupResponse'

-> ModifyAvailabilityZoneGroupResponse

Create a value of ModifyAvailabilityZoneGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:return':ModifyAvailabilityZoneGroupResponse', modifyAvailabilityZoneGroupResponse_return - Is true if the request succeeds, and an error otherwise.

$sel:httpStatus:ModifyAvailabilityZoneGroupResponse', modifyAvailabilityZoneGroupResponse_httpStatus - The response's http status code.

Response Lenses

modifyAvailabilityZoneGroupResponse_return :: Lens' ModifyAvailabilityZoneGroupResponse (Maybe Bool)

Is true if the request succeeds, and an error otherwise.

modifyAvailabilityZoneGroupResponse_httpStatus :: Lens' ModifyAvailabilityZoneGroupResponse Int

The response's http status code.