Network.AWS.EC2.ModifyDefaultCreditSpecification

Description

Modifies the default credit option for CPU usage of burstable performance instances. The default credit option is set at the account level per Amazon Web Services Region, and is specified per instance family. All new burstable performance instances in the account launch using the default credit option.

ModifyDefaultCreditSpecification is an asynchronous operation, which works at an Amazon Web Services Region level and modifies the credit option for each Availability Zone. All zones in a Region are updated within five minutes. But if instances are launched during this operation, they might not get the new credit option until the zone is updated. To verify whether the update has occurred, you can call GetDefaultCreditSpecification and check DefaultCreditSpecification for updates.

For more information, see Burstable performance instances in the Amazon EC2 User Guide.

Creating a Request

data ModifyDefaultCreditSpecification

See: newModifyDefaultCreditSpecification smart constructor.

Constructors

ModifyDefaultCreditSpecification'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

instanceFamily :: UnlimitedSupportedInstanceFamily

The instance family.

cpuCredits :: Text

The credit option for CPU usage of the instance family.

Valid Values: standard | unlimited

newModifyDefaultCreditSpecification

Arguments

:: UnlimitedSupportedInstanceFamily
$sel:instanceFamily:ModifyDefaultCreditSpecification'

-> Text
$sel:cpuCredits:ModifyDefaultCreditSpecification'

-> ModifyDefaultCreditSpecification

Create a value of ModifyDefaultCreditSpecification with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:ModifyDefaultCreditSpecification', modifyDefaultCreditSpecification_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:instanceFamily:ModifyDefaultCreditSpecification', modifyDefaultCreditSpecification_instanceFamily - The instance family.

$sel:cpuCredits:ModifyDefaultCreditSpecification', modifyDefaultCreditSpecification_cpuCredits - The credit option for CPU usage of the instance family.

Valid Values: standard | unlimited

Request Lenses

modifyDefaultCreditSpecification_dryRun :: Lens' ModifyDefaultCreditSpecification (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

modifyDefaultCreditSpecification_instanceFamily :: Lens' ModifyDefaultCreditSpecification UnlimitedSupportedInstanceFamily

The instance family.

modifyDefaultCreditSpecification_cpuCredits :: Lens' ModifyDefaultCreditSpecification Text

The credit option for CPU usage of the instance family.

Valid Values: standard | unlimited

Destructuring the Response

data ModifyDefaultCreditSpecificationResponse

See: newModifyDefaultCreditSpecificationResponse smart constructor.

Constructors

ModifyDefaultCreditSpecificationResponse'

Fields

instanceFamilyCreditSpecification :: Maybe InstanceFamilyCreditSpecification

The default credit option for CPU usage of the instance family.

httpStatus :: Int

The response's http status code.

newModifyDefaultCreditSpecificationResponse

Arguments

:: Int
$sel:httpStatus:ModifyDefaultCreditSpecificationResponse'

-> ModifyDefaultCreditSpecificationResponse

Create a value of ModifyDefaultCreditSpecificationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceFamilyCreditSpecification:ModifyDefaultCreditSpecificationResponse', modifyDefaultCreditSpecificationResponse_instanceFamilyCreditSpecification - The default credit option for CPU usage of the instance family.

$sel:httpStatus:ModifyDefaultCreditSpecificationResponse', modifyDefaultCreditSpecificationResponse_httpStatus - The response's http status code.

Response Lenses

modifyDefaultCreditSpecificationResponse_instanceFamilyCreditSpecification :: Lens' ModifyDefaultCreditSpecificationResponse (Maybe InstanceFamilyCreditSpecification)

The default credit option for CPU usage of the instance family.

modifyDefaultCreditSpecificationResponse_httpStatus :: Lens' ModifyDefaultCreditSpecificationResponse Int

The response's http status code.