Network.AWS.EC2.ModifyHosts

Description

Modify the auto-placement setting of a Dedicated Host. When auto-placement is enabled, any instances that you launch with a tenancy of host but without a specific host ID are placed onto any available Dedicated Host in your account that has auto-placement enabled. When auto-placement is disabled, you need to provide a host ID to have the instance launch onto a specific host. If no host ID is provided, the instance is launched onto a suitable host with auto-placement enabled.

You can also use this API action to modify a Dedicated Host to support either multiple instance types in an instance family, or to support a specific instance type only.

Creating a Request

data ModifyHosts

See: newModifyHosts smart constructor.

Constructors

ModifyHosts'

Fields

instanceFamily :: Maybe Text

Specifies the instance family to be supported by the Dedicated Host. Specify this parameter to modify a Dedicated Host to support multiple instance types within its current instance family.

If you want to modify a Dedicated Host to support a specific instance type only, omit this parameter and specify InstanceType instead. You cannot specify InstanceFamily and InstanceType in the same request.

instanceType :: Maybe Text

Specifies the instance type to be supported by the Dedicated Host. Specify this parameter to modify a Dedicated Host to support only a specific instance type.

If you want to modify a Dedicated Host to support multiple instance types in its current instance family, omit this parameter and specify InstanceFamily instead. You cannot specify InstanceType and InstanceFamily in the same request.

autoPlacement :: Maybe AutoPlacement

Specify whether to enable or disable auto-placement.

hostRecovery :: Maybe HostRecovery

Indicates whether to enable or disable host recovery for the Dedicated Host. For more information, see Host recovery in the Amazon EC2 User Guide.

hostIds :: [Text]

The IDs of the Dedicated Hosts to modify.

newModifyHosts :: ModifyHosts

Create a value of ModifyHosts with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceFamily:ModifyHosts', modifyHosts_instanceFamily - Specifies the instance family to be supported by the Dedicated Host. Specify this parameter to modify a Dedicated Host to support multiple instance types within its current instance family.

If you want to modify a Dedicated Host to support a specific instance type only, omit this parameter and specify InstanceType instead. You cannot specify InstanceFamily and InstanceType in the same request.

$sel:instanceType:ModifyHosts', modifyHosts_instanceType - Specifies the instance type to be supported by the Dedicated Host. Specify this parameter to modify a Dedicated Host to support only a specific instance type.

If you want to modify a Dedicated Host to support multiple instance types in its current instance family, omit this parameter and specify InstanceFamily instead. You cannot specify InstanceType and InstanceFamily in the same request.

$sel:autoPlacement:ModifyHosts', modifyHosts_autoPlacement - Specify whether to enable or disable auto-placement.

$sel:hostRecovery:ModifyHosts', modifyHosts_hostRecovery - Indicates whether to enable or disable host recovery for the Dedicated Host. For more information, see Host recovery in the Amazon EC2 User Guide.

$sel:hostIds:ModifyHosts', modifyHosts_hostIds - The IDs of the Dedicated Hosts to modify.

Request Lenses

modifyHosts_instanceFamily :: Lens' ModifyHosts (Maybe Text)

Specifies the instance family to be supported by the Dedicated Host. Specify this parameter to modify a Dedicated Host to support multiple instance types within its current instance family.

If you want to modify a Dedicated Host to support a specific instance type only, omit this parameter and specify InstanceType instead. You cannot specify InstanceFamily and InstanceType in the same request.

modifyHosts_instanceType :: Lens' ModifyHosts (Maybe Text)

Specifies the instance type to be supported by the Dedicated Host. Specify this parameter to modify a Dedicated Host to support only a specific instance type.

If you want to modify a Dedicated Host to support multiple instance types in its current instance family, omit this parameter and specify InstanceFamily instead. You cannot specify InstanceType and InstanceFamily in the same request.

modifyHosts_autoPlacement :: Lens' ModifyHosts (Maybe AutoPlacement)

Specify whether to enable or disable auto-placement.

modifyHosts_hostRecovery :: Lens' ModifyHosts (Maybe HostRecovery)

Indicates whether to enable or disable host recovery for the Dedicated Host. For more information, see Host recovery in the Amazon EC2 User Guide.

modifyHosts_hostIds :: Lens' ModifyHosts [Text]

The IDs of the Dedicated Hosts to modify.

Destructuring the Response

data ModifyHostsResponse

See: newModifyHostsResponse smart constructor.

Constructors

ModifyHostsResponse'

Fields

unsuccessful :: Maybe [UnsuccessfulItem]

The IDs of the Dedicated Hosts that could not be modified. Check whether the setting you requested can be used.

successful :: Maybe [Text]

The IDs of the Dedicated Hosts that were successfully modified.

httpStatus :: Int

The response's http status code.

newModifyHostsResponse

Arguments

:: Int
$sel:httpStatus:ModifyHostsResponse'

-> ModifyHostsResponse

Create a value of ModifyHostsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:unsuccessful:ModifyHostsResponse', modifyHostsResponse_unsuccessful - The IDs of the Dedicated Hosts that could not be modified. Check whether the setting you requested can be used.

$sel:successful:ModifyHostsResponse', modifyHostsResponse_successful - The IDs of the Dedicated Hosts that were successfully modified.

$sel:httpStatus:ModifyHostsResponse', modifyHostsResponse_httpStatus - The response's http status code.

Response Lenses

modifyHostsResponse_unsuccessful :: Lens' ModifyHostsResponse (Maybe [UnsuccessfulItem])

The IDs of the Dedicated Hosts that could not be modified. Check whether the setting you requested can be used.

modifyHostsResponse_successful :: Lens' ModifyHostsResponse (Maybe [Text])

The IDs of the Dedicated Hosts that were successfully modified.

modifyHostsResponse_httpStatus :: Lens' ModifyHostsResponse Int

The response's http status code.