Network.AWS.EC2.ModifyIdFormat

Description

Modifies the ID format for the specified resource on a per-Region basis. You can specify that resources should receive longer IDs (17-character IDs) when they are created.

This request can only be used to modify longer ID settings for resource types that are within the opt-in period. Resources currently in their opt-in period include: bundle | conversion-task | customer-gateway | dhcp-options | elastic-ip-allocation | elastic-ip-association | export-task | flow-log | image | import-task | internet-gateway | network-acl | network-acl-association | network-interface | network-interface-attachment | prefix-list | route-table | route-table-association | security-group | subnet | subnet-cidr-block-association | vpc | vpc-cidr-block-association | vpc-endpoint | vpc-peering-connection | vpn-connection | vpn-gateway.

This setting applies to the IAM user who makes the request; it does not apply to the entire AWS account. By default, an IAM user defaults to the same settings as the root user. If you're using this action as the root user, then these settings apply to the entire account, unless an IAM user explicitly overrides these settings for themselves. For more information, see Resource IDs in the Amazon Elastic Compute Cloud User Guide.

Resources created with longer IDs are visible to all IAM roles and users, regardless of these settings and provided that they have permission to use the relevant Describe command for the resource type.

Creating a Request

data ModifyIdFormat

See: newModifyIdFormat smart constructor.

Constructors

ModifyIdFormat'

Fields

resource :: Text

The type of resource: bundle | conversion-task | customer-gateway | dhcp-options | elastic-ip-allocation | elastic-ip-association | export-task | flow-log | image | import-task | internet-gateway | network-acl | network-acl-association | network-interface | network-interface-attachment | prefix-list | route-table | route-table-association | security-group | subnet | subnet-cidr-block-association | vpc | vpc-cidr-block-association | vpc-endpoint | vpc-peering-connection | vpn-connection | vpn-gateway.

Alternatively, use the all-current option to include all resource types that are currently within their opt-in period for longer IDs.

useLongIds :: Bool

Indicate whether the resource should use longer IDs (17-character IDs).

newModifyIdFormat

Arguments

:: Text
$sel:resource:ModifyIdFormat'

-> Bool
$sel:useLongIds:ModifyIdFormat'

-> ModifyIdFormat

Create a value of ModifyIdFormat with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resource:ModifyIdFormat', modifyIdFormat_resource - The type of resource: bundle | conversion-task | customer-gateway | dhcp-options | elastic-ip-allocation | elastic-ip-association | export-task | flow-log | image | import-task | internet-gateway | network-acl | network-acl-association | network-interface | network-interface-attachment | prefix-list | route-table | route-table-association | security-group | subnet | subnet-cidr-block-association | vpc | vpc-cidr-block-association | vpc-endpoint | vpc-peering-connection | vpn-connection | vpn-gateway.

Alternatively, use the all-current option to include all resource types that are currently within their opt-in period for longer IDs.

$sel:useLongIds:ModifyIdFormat', modifyIdFormat_useLongIds - Indicate whether the resource should use longer IDs (17-character IDs).

Request Lenses

modifyIdFormat_resource :: Lens' ModifyIdFormat Text

The type of resource: bundle | conversion-task | customer-gateway | dhcp-options | elastic-ip-allocation | elastic-ip-association | export-task | flow-log | image | import-task | internet-gateway | network-acl | network-acl-association | network-interface | network-interface-attachment | prefix-list | route-table | route-table-association | security-group | subnet | subnet-cidr-block-association | vpc | vpc-cidr-block-association | vpc-endpoint | vpc-peering-connection | vpn-connection | vpn-gateway.

Alternatively, use the all-current option to include all resource types that are currently within their opt-in period for longer IDs.

modifyIdFormat_useLongIds :: Lens' ModifyIdFormat Bool

Indicate whether the resource should use longer IDs (17-character IDs).

Destructuring the Response

data ModifyIdFormatResponse

See: newModifyIdFormatResponse smart constructor.

Constructors

ModifyIdFormatResponse'

newModifyIdFormatResponse :: ModifyIdFormatResponse

Create a value of ModifyIdFormatResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.