Network.AWS.EC2.ModifyInstanceEventWindow

Description

Modifies the specified event window.

You can define either a set of time ranges or a cron expression when modifying the event window, but not both.

To modify the targets associated with the event window, use the AssociateInstanceEventWindow and DisassociateInstanceEventWindow API.

If Amazon Web Services has already scheduled an event, modifying an event window won't change the time of the scheduled event.

For more information, see Define event windows for scheduled events in the Amazon EC2 User Guide.

Creating a Request

data ModifyInstanceEventWindow

See: newModifyInstanceEventWindow smart constructor.

Constructors

ModifyInstanceEventWindow'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

name :: Maybe Text

The name of the event window.

timeRanges :: Maybe [InstanceEventWindowTimeRangeRequest]

The time ranges of the event window.

cronExpression :: Maybe Text

The cron expression of the event window, for example, * 0-4,20-23 * * 1,5.

Constraints:

Only hour and day of the week values are supported.

For day of the week values, you can specify either integers 0 through 6, or alternative single values SUN through SAT.

The minute, month, and year must be specified by *.

The hour value must be one or a multiple range, for example, 0-4 or 0-4,20-23.

Each hour range must be >= 2 hours, for example, 0-2 or 20-23.

The event window must be >= 4 hours. The combined total time ranges in the event window must be >= 4 hours.

For more information about cron expressions, see cron on the Wikipedia website.

instanceEventWindowId :: Text

The ID of the event window.

newModifyInstanceEventWindow

Arguments

:: Text
$sel:instanceEventWindowId:ModifyInstanceEventWindow'

-> ModifyInstanceEventWindow

Create a value of ModifyInstanceEventWindow with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:ModifyInstanceEventWindow', modifyInstanceEventWindow_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:name:ModifyInstanceEventWindow', modifyInstanceEventWindow_name - The name of the event window.

$sel:timeRanges:ModifyInstanceEventWindow', modifyInstanceEventWindow_timeRanges - The time ranges of the event window.

$sel:cronExpression:ModifyInstanceEventWindow', modifyInstanceEventWindow_cronExpression - The cron expression of the event window, for example, * 0-4,20-23 * * 1,5.

Constraints:

Only hour and day of the week values are supported.

For day of the week values, you can specify either integers 0 through 6, or alternative single values SUN through SAT.

The minute, month, and year must be specified by *.

The hour value must be one or a multiple range, for example, 0-4 or 0-4,20-23.

Each hour range must be >= 2 hours, for example, 0-2 or 20-23.

The event window must be >= 4 hours. The combined total time ranges in the event window must be >= 4 hours.

For more information about cron expressions, see cron on the Wikipedia website.

$sel:instanceEventWindowId:ModifyInstanceEventWindow', modifyInstanceEventWindow_instanceEventWindowId - The ID of the event window.

Request Lenses

modifyInstanceEventWindow_dryRun :: Lens' ModifyInstanceEventWindow (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

modifyInstanceEventWindow_name :: Lens' ModifyInstanceEventWindow (Maybe Text)

The name of the event window.

modifyInstanceEventWindow_timeRanges :: Lens' ModifyInstanceEventWindow (Maybe [InstanceEventWindowTimeRangeRequest])

The time ranges of the event window.

modifyInstanceEventWindow_cronExpression :: Lens' ModifyInstanceEventWindow (Maybe Text)

The cron expression of the event window, for example, * 0-4,20-23 * * 1,5.

Constraints:

Only hour and day of the week values are supported.

For day of the week values, you can specify either integers 0 through 6, or alternative single values SUN through SAT.

The minute, month, and year must be specified by *.

The hour value must be one or a multiple range, for example, 0-4 or 0-4,20-23.

Each hour range must be >= 2 hours, for example, 0-2 or 20-23.

The event window must be >= 4 hours. The combined total time ranges in the event window must be >= 4 hours.

For more information about cron expressions, see cron on the Wikipedia website.

modifyInstanceEventWindow_instanceEventWindowId :: Lens' ModifyInstanceEventWindow Text

The ID of the event window.

Destructuring the Response

data ModifyInstanceEventWindowResponse

See: newModifyInstanceEventWindowResponse smart constructor.

Constructors

ModifyInstanceEventWindowResponse'

Fields

instanceEventWindow :: Maybe InstanceEventWindow

Information about the event window.

httpStatus :: Int

The response's http status code.

newModifyInstanceEventWindowResponse

Arguments

:: Int
$sel:httpStatus:ModifyInstanceEventWindowResponse'

-> ModifyInstanceEventWindowResponse

Create a value of ModifyInstanceEventWindowResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceEventWindow:ModifyInstanceEventWindowResponse', modifyInstanceEventWindowResponse_instanceEventWindow - Information about the event window.

$sel:httpStatus:ModifyInstanceEventWindowResponse', modifyInstanceEventWindowResponse_httpStatus - The response's http status code.

Response Lenses

modifyInstanceEventWindowResponse_instanceEventWindow :: Lens' ModifyInstanceEventWindowResponse (Maybe InstanceEventWindow)

Information about the event window.

modifyInstanceEventWindowResponse_httpStatus :: Lens' ModifyInstanceEventWindowResponse Int

The response's http status code.