Network.AWS.EC2.ModifyInstanceMetadataOptions

Description

Modify the instance metadata parameters on a running or stopped instance. When you modify the parameters on a stopped instance, they are applied when the instance is started. When you modify the parameters on a running instance, the API responds with a state of “pending”. After the parameter modifications are successfully applied to the instance, the state of the modifications changes from “pending” to “applied” in subsequent describe-instances API calls. For more information, see Instance metadata and user data in the Amazon EC2 User Guide.

Creating a Request

data ModifyInstanceMetadataOptions

See: newModifyInstanceMetadataOptions smart constructor.

Constructors

ModifyInstanceMetadataOptions'

Fields

httpProtocolIpv6 :: Maybe InstanceMetadataProtocolState

Enables or disables the IPv6 endpoint for the instance metadata service.

httpEndpoint :: Maybe InstanceMetadataEndpointState

This parameter enables or disables the HTTP metadata endpoint on your instances. If the parameter is not specified, the existing state is maintained.

If you specify a value of disabled, you will not be able to access your instance metadata.

httpPutResponseHopLimit :: Maybe Int

The desired HTTP PUT response hop limit for instance metadata requests. The larger the number, the further instance metadata requests can travel. If no parameter is specified, the existing state is maintained.

Possible values: Integers from 1 to 64

httpTokens :: Maybe HttpTokensState

The state of token usage for your instance metadata requests. If the parameter is not specified in the request, the default state is optional.

If the state is optional, you can choose to retrieve instance metadata with or without a signed token header on your request. If you retrieve the IAM role credentials without a token, the version 1.0 role credentials are returned. If you retrieve the IAM role credentials using a valid signed token, the version 2.0 role credentials are returned.

If the state is required, you must send a signed token header with any instance metadata retrieval requests. In this state, retrieving the IAM role credential always returns the version 2.0 credentials; the version 1.0 credentials are not available.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

instanceId :: Text

The ID of the instance.

newModifyInstanceMetadataOptions

Arguments

:: Text
$sel:instanceId:ModifyInstanceMetadataOptions'

-> ModifyInstanceMetadataOptions

Create a value of ModifyInstanceMetadataOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpProtocolIpv6:ModifyInstanceMetadataOptions', modifyInstanceMetadataOptions_httpProtocolIpv6 - Enables or disables the IPv6 endpoint for the instance metadata service.

$sel:httpEndpoint:ModifyInstanceMetadataOptions', modifyInstanceMetadataOptions_httpEndpoint - This parameter enables or disables the HTTP metadata endpoint on your instances. If the parameter is not specified, the existing state is maintained.

If you specify a value of disabled, you will not be able to access your instance metadata.

$sel:httpPutResponseHopLimit:ModifyInstanceMetadataOptions', modifyInstanceMetadataOptions_httpPutResponseHopLimit - The desired HTTP PUT response hop limit for instance metadata requests. The larger the number, the further instance metadata requests can travel. If no parameter is specified, the existing state is maintained.

Possible values: Integers from 1 to 64

$sel:httpTokens:ModifyInstanceMetadataOptions', modifyInstanceMetadataOptions_httpTokens - The state of token usage for your instance metadata requests. If the parameter is not specified in the request, the default state is optional.

If the state is optional, you can choose to retrieve instance metadata with or without a signed token header on your request. If you retrieve the IAM role credentials without a token, the version 1.0 role credentials are returned. If you retrieve the IAM role credentials using a valid signed token, the version 2.0 role credentials are returned.

If the state is required, you must send a signed token header with any instance metadata retrieval requests. In this state, retrieving the IAM role credential always returns the version 2.0 credentials; the version 1.0 credentials are not available.

$sel:dryRun:ModifyInstanceMetadataOptions', modifyInstanceMetadataOptions_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:instanceId:ModifyInstanceMetadataOptions', modifyInstanceMetadataOptions_instanceId - The ID of the instance.

Request Lenses

modifyInstanceMetadataOptions_httpProtocolIpv6 :: Lens' ModifyInstanceMetadataOptions (Maybe InstanceMetadataProtocolState)

Enables or disables the IPv6 endpoint for the instance metadata service.

modifyInstanceMetadataOptions_httpEndpoint :: Lens' ModifyInstanceMetadataOptions (Maybe InstanceMetadataEndpointState)

This parameter enables or disables the HTTP metadata endpoint on your instances. If the parameter is not specified, the existing state is maintained.

If you specify a value of disabled, you will not be able to access your instance metadata.

modifyInstanceMetadataOptions_httpPutResponseHopLimit :: Lens' ModifyInstanceMetadataOptions (Maybe Int)

The desired HTTP PUT response hop limit for instance metadata requests. The larger the number, the further instance metadata requests can travel. If no parameter is specified, the existing state is maintained.

Possible values: Integers from 1 to 64

modifyInstanceMetadataOptions_httpTokens :: Lens' ModifyInstanceMetadataOptions (Maybe HttpTokensState)

The state of token usage for your instance metadata requests. If the parameter is not specified in the request, the default state is optional.

If the state is optional, you can choose to retrieve instance metadata with or without a signed token header on your request. If you retrieve the IAM role credentials without a token, the version 1.0 role credentials are returned. If you retrieve the IAM role credentials using a valid signed token, the version 2.0 role credentials are returned.

If the state is required, you must send a signed token header with any instance metadata retrieval requests. In this state, retrieving the IAM role credential always returns the version 2.0 credentials; the version 1.0 credentials are not available.

modifyInstanceMetadataOptions_dryRun :: Lens' ModifyInstanceMetadataOptions (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

modifyInstanceMetadataOptions_instanceId :: Lens' ModifyInstanceMetadataOptions Text

The ID of the instance.

Destructuring the Response

data ModifyInstanceMetadataOptionsResponse

See: newModifyInstanceMetadataOptionsResponse smart constructor.

Constructors

ModifyInstanceMetadataOptionsResponse'

Fields

instanceId :: Maybe Text

The ID of the instance.

instanceMetadataOptions :: Maybe InstanceMetadataOptionsResponse

The metadata options for the instance.

httpStatus :: Int

The response's http status code.

newModifyInstanceMetadataOptionsResponse

Arguments

:: Int
$sel:httpStatus:ModifyInstanceMetadataOptionsResponse'

-> ModifyInstanceMetadataOptionsResponse

Create a value of ModifyInstanceMetadataOptionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceId:ModifyInstanceMetadataOptions', modifyInstanceMetadataOptionsResponse_instanceId - The ID of the instance.

$sel:instanceMetadataOptions:ModifyInstanceMetadataOptionsResponse', modifyInstanceMetadataOptionsResponse_instanceMetadataOptions - The metadata options for the instance.

$sel:httpStatus:ModifyInstanceMetadataOptionsResponse', modifyInstanceMetadataOptionsResponse_httpStatus - The response's http status code.

Response Lenses

modifyInstanceMetadataOptionsResponse_instanceId :: Lens' ModifyInstanceMetadataOptionsResponse (Maybe Text)

The ID of the instance.

modifyInstanceMetadataOptionsResponse_instanceMetadataOptions :: Lens' ModifyInstanceMetadataOptionsResponse (Maybe InstanceMetadataOptionsResponse)

The metadata options for the instance.

modifyInstanceMetadataOptionsResponse_httpStatus :: Lens' ModifyInstanceMetadataOptionsResponse Int

The response's http status code.