Network.AWS.EC2.ModifyLaunchTemplate

Description

Modifies a launch template. You can specify which version of the launch template to set as the default version. When launching an instance, the default version applies when a launch template version is not specified.

Creating a Request

data ModifyLaunchTemplate

See: newModifyLaunchTemplate smart constructor.

Constructors

ModifyLaunchTemplate'

Fields

defaultVersion :: Maybe Text

The version number of the launch template to set as the default version.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

launchTemplateId :: Maybe Text

The ID of the launch template. You must specify either the launch template ID or launch template name in the request.

launchTemplateName :: Maybe Text

The name of the launch template. You must specify either the launch template ID or launch template name in the request.

clientToken :: Maybe Text

Unique, case-sensitive identifier you provide to ensure the idempotency of the request. For more information, see Ensuring Idempotency.

Constraint: Maximum 128 ASCII characters.

newModifyLaunchTemplate :: ModifyLaunchTemplate

Create a value of ModifyLaunchTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:defaultVersion:ModifyLaunchTemplate', modifyLaunchTemplate_defaultVersion - The version number of the launch template to set as the default version.

$sel:dryRun:ModifyLaunchTemplate', modifyLaunchTemplate_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:launchTemplateId:ModifyLaunchTemplate', modifyLaunchTemplate_launchTemplateId - The ID of the launch template. You must specify either the launch template ID or launch template name in the request.

$sel:launchTemplateName:ModifyLaunchTemplate', modifyLaunchTemplate_launchTemplateName - The name of the launch template. You must specify either the launch template ID or launch template name in the request.

$sel:clientToken:ModifyLaunchTemplate', modifyLaunchTemplate_clientToken - Unique, case-sensitive identifier you provide to ensure the idempotency of the request. For more information, see Ensuring Idempotency.

Constraint: Maximum 128 ASCII characters.

Request Lenses

modifyLaunchTemplate_defaultVersion :: Lens' ModifyLaunchTemplate (Maybe Text)

The version number of the launch template to set as the default version.

modifyLaunchTemplate_dryRun :: Lens' ModifyLaunchTemplate (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

modifyLaunchTemplate_launchTemplateId :: Lens' ModifyLaunchTemplate (Maybe Text)

The ID of the launch template. You must specify either the launch template ID or launch template name in the request.

modifyLaunchTemplate_launchTemplateName :: Lens' ModifyLaunchTemplate (Maybe Text)

The name of the launch template. You must specify either the launch template ID or launch template name in the request.

modifyLaunchTemplate_clientToken :: Lens' ModifyLaunchTemplate (Maybe Text)

Unique, case-sensitive identifier you provide to ensure the idempotency of the request. For more information, see Ensuring Idempotency.

Constraint: Maximum 128 ASCII characters.

Destructuring the Response

data ModifyLaunchTemplateResponse

See: newModifyLaunchTemplateResponse smart constructor.

Constructors

ModifyLaunchTemplateResponse'

Fields

launchTemplate :: Maybe LaunchTemplate

Information about the launch template.

httpStatus :: Int

The response's http status code.

newModifyLaunchTemplateResponse

Arguments

:: Int
$sel:httpStatus:ModifyLaunchTemplateResponse'

-> ModifyLaunchTemplateResponse

Create a value of ModifyLaunchTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:launchTemplate:ModifyLaunchTemplateResponse', modifyLaunchTemplateResponse_launchTemplate - Information about the launch template.

$sel:httpStatus:ModifyLaunchTemplateResponse', modifyLaunchTemplateResponse_httpStatus - The response's http status code.

Response Lenses

modifyLaunchTemplateResponse_launchTemplate :: Lens' ModifyLaunchTemplateResponse (Maybe LaunchTemplate)

Information about the launch template.

modifyLaunchTemplateResponse_httpStatus :: Lens' ModifyLaunchTemplateResponse Int

The response's http status code.