Network.AWS.EC2.ModifyManagedPrefixList

Description

Modifies the specified managed prefix list.

Adding or removing entries in a prefix list creates a new version of the prefix list. Changing the name of the prefix list does not affect the version.

If you specify a current version number that does not match the true current version number, the request fails.

Creating a Request

data ModifyManagedPrefixList

See: newModifyManagedPrefixList smart constructor.

Constructors

ModifyManagedPrefixList'

Fields

currentVersion :: Maybe Integer

The current version of the prefix list.

removeEntries :: Maybe [RemovePrefixListEntry]

One or more entries to remove from the prefix list.

prefixListName :: Maybe Text

A name for the prefix list.

addEntries :: Maybe [AddPrefixListEntry]

One or more entries to add to the prefix list.

maxEntries :: Maybe Int

The maximum number of entries for the prefix list. You cannot modify the entries of a prefix list and modify the size of a prefix list at the same time.

If any of the resources that reference the prefix list cannot support the new maximum size, the modify operation fails. Check the state message for the IDs of the first ten resources that do not support the new maximum size.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

prefixListId :: Text

The ID of the prefix list.

newModifyManagedPrefixList

Arguments

:: Text
$sel:prefixListId:ModifyManagedPrefixList'

-> ModifyManagedPrefixList

Create a value of ModifyManagedPrefixList with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:currentVersion:ModifyManagedPrefixList', modifyManagedPrefixList_currentVersion - The current version of the prefix list.

$sel:removeEntries:ModifyManagedPrefixList', modifyManagedPrefixList_removeEntries - One or more entries to remove from the prefix list.

$sel:prefixListName:ModifyManagedPrefixList', modifyManagedPrefixList_prefixListName - A name for the prefix list.

$sel:addEntries:ModifyManagedPrefixList', modifyManagedPrefixList_addEntries - One or more entries to add to the prefix list.

$sel:maxEntries:ModifyManagedPrefixList', modifyManagedPrefixList_maxEntries - The maximum number of entries for the prefix list. You cannot modify the entries of a prefix list and modify the size of a prefix list at the same time.

If any of the resources that reference the prefix list cannot support the new maximum size, the modify operation fails. Check the state message for the IDs of the first ten resources that do not support the new maximum size.

$sel:dryRun:ModifyManagedPrefixList', modifyManagedPrefixList_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:prefixListId:ModifyManagedPrefixList', modifyManagedPrefixList_prefixListId - The ID of the prefix list.

Request Lenses

modifyManagedPrefixList_currentVersion :: Lens' ModifyManagedPrefixList (Maybe Integer)

The current version of the prefix list.

modifyManagedPrefixList_removeEntries :: Lens' ModifyManagedPrefixList (Maybe [RemovePrefixListEntry])

One or more entries to remove from the prefix list.

modifyManagedPrefixList_prefixListName :: Lens' ModifyManagedPrefixList (Maybe Text)

A name for the prefix list.

modifyManagedPrefixList_addEntries :: Lens' ModifyManagedPrefixList (Maybe [AddPrefixListEntry])

One or more entries to add to the prefix list.

modifyManagedPrefixList_maxEntries :: Lens' ModifyManagedPrefixList (Maybe Int)

The maximum number of entries for the prefix list. You cannot modify the entries of a prefix list and modify the size of a prefix list at the same time.

If any of the resources that reference the prefix list cannot support the new maximum size, the modify operation fails. Check the state message for the IDs of the first ten resources that do not support the new maximum size.

modifyManagedPrefixList_dryRun :: Lens' ModifyManagedPrefixList (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

modifyManagedPrefixList_prefixListId :: Lens' ModifyManagedPrefixList Text

The ID of the prefix list.

Destructuring the Response

data ModifyManagedPrefixListResponse

See: newModifyManagedPrefixListResponse smart constructor.

Constructors

ModifyManagedPrefixListResponse'

Fields

prefixList :: Maybe ManagedPrefixList

Information about the prefix list.

httpStatus :: Int

The response's http status code.

newModifyManagedPrefixListResponse

Arguments

:: Int
$sel:httpStatus:ModifyManagedPrefixListResponse'

-> ModifyManagedPrefixListResponse

Create a value of ModifyManagedPrefixListResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:prefixList:ModifyManagedPrefixListResponse', modifyManagedPrefixListResponse_prefixList - Information about the prefix list.

$sel:httpStatus:ModifyManagedPrefixListResponse', modifyManagedPrefixListResponse_httpStatus - The response's http status code.

Response Lenses

modifyManagedPrefixListResponse_prefixList :: Lens' ModifyManagedPrefixListResponse (Maybe ManagedPrefixList)

Information about the prefix list.

modifyManagedPrefixListResponse_httpStatus :: Lens' ModifyManagedPrefixListResponse Int

The response's http status code.