Network.AWS.EC2.ModifyNetworkInterfaceAttribute

Description

Modifies the specified network interface attribute. You can specify only one attribute at a time. You can use this action to attach and detach security groups from an existing EC2 instance.

Creating a Request

data ModifyNetworkInterfaceAttribute

Contains the parameters for ModifyNetworkInterfaceAttribute.

See: newModifyNetworkInterfaceAttribute smart constructor.

Constructors

ModifyNetworkInterfaceAttribute'

Fields

groups :: Maybe [Text]

Changes the security groups for the network interface. The new set of groups you specify replaces the current set. You must specify at least one group, even if it's just the default security group in the VPC. You must specify the ID of the security group, not the name.

attachment :: Maybe NetworkInterfaceAttachmentChanges

Information about the interface attachment. If modifying the 'delete on termination' attribute, you must specify the ID of the interface attachment.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

sourceDestCheck :: Maybe AttributeBooleanValue

Enable or disable source/destination checks, which ensure that the instance is either the source or the destination of any traffic that it receives. If the value is true, source/destination checks are enabled; otherwise, they are disabled. The default value is true. You must disable source/destination checks if the instance runs services such as network address translation, routing, or firewalls.

description :: Maybe AttributeValue

A description for the network interface.

networkInterfaceId :: Text

The ID of the network interface.

newModifyNetworkInterfaceAttribute

Arguments

:: Text
$sel:networkInterfaceId:ModifyNetworkInterfaceAttribute'

-> ModifyNetworkInterfaceAttribute

Create a value of ModifyNetworkInterfaceAttribute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groups:ModifyNetworkInterfaceAttribute', modifyNetworkInterfaceAttribute_groups - Changes the security groups for the network interface. The new set of groups you specify replaces the current set. You must specify at least one group, even if it's just the default security group in the VPC. You must specify the ID of the security group, not the name.

$sel:attachment:ModifyNetworkInterfaceAttribute', modifyNetworkInterfaceAttribute_attachment - Information about the interface attachment. If modifying the 'delete on termination' attribute, you must specify the ID of the interface attachment.

$sel:dryRun:ModifyNetworkInterfaceAttribute', modifyNetworkInterfaceAttribute_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:sourceDestCheck:ModifyNetworkInterfaceAttribute', modifyNetworkInterfaceAttribute_sourceDestCheck - Enable or disable source/destination checks, which ensure that the instance is either the source or the destination of any traffic that it receives. If the value is true, source/destination checks are enabled; otherwise, they are disabled. The default value is true. You must disable source/destination checks if the instance runs services such as network address translation, routing, or firewalls.

$sel:description:ModifyNetworkInterfaceAttribute', modifyNetworkInterfaceAttribute_description - A description for the network interface.

$sel:networkInterfaceId:ModifyNetworkInterfaceAttribute', modifyNetworkInterfaceAttribute_networkInterfaceId - The ID of the network interface.

Request Lenses

modifyNetworkInterfaceAttribute_groups :: Lens' ModifyNetworkInterfaceAttribute (Maybe [Text])

Changes the security groups for the network interface. The new set of groups you specify replaces the current set. You must specify at least one group, even if it's just the default security group in the VPC. You must specify the ID of the security group, not the name.

modifyNetworkInterfaceAttribute_attachment :: Lens' ModifyNetworkInterfaceAttribute (Maybe NetworkInterfaceAttachmentChanges)

Information about the interface attachment. If modifying the 'delete on termination' attribute, you must specify the ID of the interface attachment.

modifyNetworkInterfaceAttribute_dryRun :: Lens' ModifyNetworkInterfaceAttribute (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

modifyNetworkInterfaceAttribute_sourceDestCheck :: Lens' ModifyNetworkInterfaceAttribute (Maybe AttributeBooleanValue)

Enable or disable source/destination checks, which ensure that the instance is either the source or the destination of any traffic that it receives. If the value is true, source/destination checks are enabled; otherwise, they are disabled. The default value is true. You must disable source/destination checks if the instance runs services such as network address translation, routing, or firewalls.

modifyNetworkInterfaceAttribute_description :: Lens' ModifyNetworkInterfaceAttribute (Maybe AttributeValue)

A description for the network interface.

modifyNetworkInterfaceAttribute_networkInterfaceId :: Lens' ModifyNetworkInterfaceAttribute Text

The ID of the network interface.

Destructuring the Response

data ModifyNetworkInterfaceAttributeResponse

See: newModifyNetworkInterfaceAttributeResponse smart constructor.

Constructors

ModifyNetworkInterfaceAttributeResponse'

newModifyNetworkInterfaceAttributeResponse :: ModifyNetworkInterfaceAttributeResponse

Create a value of ModifyNetworkInterfaceAttributeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.