Network.AWS.EC2.ModifyReservedInstances

Description

Modifies the Availability Zone, instance count, instance type, or network platform (EC2-Classic or EC2-VPC) of your Reserved Instances. The Reserved Instances to be modified must be identical, except for Availability Zone, network platform, and instance type.

For more information, see Modifying Reserved Instances in the Amazon EC2 User Guide.

Creating a Request

data ModifyReservedInstances

Contains the parameters for ModifyReservedInstances.

See: newModifyReservedInstances smart constructor.

Constructors

ModifyReservedInstances'

Fields

clientToken :: Maybe Text

A unique, case-sensitive token you provide to ensure idempotency of your modification request. For more information, see Ensuring Idempotency.

reservedInstancesIds :: [Text]

The IDs of the Reserved Instances to modify.

targetConfigurations :: [ReservedInstancesConfiguration]

The configuration settings for the Reserved Instances to modify.

newModifyReservedInstances :: ModifyReservedInstances

Create a value of ModifyReservedInstances with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:ModifyReservedInstances', modifyReservedInstances_clientToken - A unique, case-sensitive token you provide to ensure idempotency of your modification request. For more information, see Ensuring Idempotency.

$sel:reservedInstancesIds:ModifyReservedInstances', modifyReservedInstances_reservedInstancesIds - The IDs of the Reserved Instances to modify.

$sel:targetConfigurations:ModifyReservedInstances', modifyReservedInstances_targetConfigurations - The configuration settings for the Reserved Instances to modify.

Request Lenses

modifyReservedInstances_clientToken :: Lens' ModifyReservedInstances (Maybe Text)

A unique, case-sensitive token you provide to ensure idempotency of your modification request. For more information, see Ensuring Idempotency.

modifyReservedInstances_reservedInstancesIds :: Lens' ModifyReservedInstances [Text]

The IDs of the Reserved Instances to modify.

modifyReservedInstances_targetConfigurations :: Lens' ModifyReservedInstances [ReservedInstancesConfiguration]

The configuration settings for the Reserved Instances to modify.

Destructuring the Response

data ModifyReservedInstancesResponse

Contains the output of ModifyReservedInstances.

See: newModifyReservedInstancesResponse smart constructor.

Constructors

ModifyReservedInstancesResponse'

Fields

reservedInstancesModificationId :: Maybe Text

The ID for the modification.

httpStatus :: Int

The response's http status code.

newModifyReservedInstancesResponse

Arguments

:: Int
$sel:httpStatus:ModifyReservedInstancesResponse'

-> ModifyReservedInstancesResponse

Create a value of ModifyReservedInstancesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reservedInstancesModificationId:ModifyReservedInstancesResponse', modifyReservedInstancesResponse_reservedInstancesModificationId - The ID for the modification.

$sel:httpStatus:ModifyReservedInstancesResponse', modifyReservedInstancesResponse_httpStatus - The response's http status code.

Response Lenses

modifyReservedInstancesResponse_reservedInstancesModificationId :: Lens' ModifyReservedInstancesResponse (Maybe Text)

The ID for the modification.

modifyReservedInstancesResponse_httpStatus :: Lens' ModifyReservedInstancesResponse Int

The response's http status code.