Network.AWS.EC2.ModifySpotFleetRequest

Description

Modifies the specified Spot Fleet request.

You can only modify a Spot Fleet request of type maintain.

While the Spot Fleet request is being modified, it is in the modifying state.

To scale up your Spot Fleet, increase its target capacity. The Spot Fleet launches the additional Spot Instances according to the allocation strategy for the Spot Fleet request. If the allocation strategy is lowestPrice, the Spot Fleet launches instances using the Spot Instance pool with the lowest price. If the allocation strategy is diversified, the Spot Fleet distributes the instances across the Spot Instance pools. If the allocation strategy is capacityOptimized, Spot Fleet launches instances from Spot Instance pools with optimal capacity for the number of instances that are launching.

To scale down your Spot Fleet, decrease its target capacity. First, the Spot Fleet cancels any open requests that exceed the new target capacity. You can request that the Spot Fleet terminate Spot Instances until the size of the fleet no longer exceeds the new target capacity. If the allocation strategy is lowestPrice, the Spot Fleet terminates the instances with the highest price per unit. If the allocation strategy is capacityOptimized, the Spot Fleet terminates the instances in the Spot Instance pools that have the least available Spot Instance capacity. If the allocation strategy is diversified, the Spot Fleet terminates instances across the Spot Instance pools. Alternatively, you can request that the Spot Fleet keep the fleet at its current size, but not replace any Spot Instances that are interrupted or that you terminate manually.

If you are finished with your Spot Fleet for now, but will use it again later, you can set the target capacity to 0.

Creating a Request

data ModifySpotFleetRequest

Contains the parameters for ModifySpotFleetRequest.

See: newModifySpotFleetRequest smart constructor.

Constructors

ModifySpotFleetRequest'

Fields

launchTemplateConfigs :: Maybe [LaunchTemplateConfig]

The launch template and overrides. You can only use this parameter if you specified a launch template (LaunchTemplateConfigs) in your Spot Fleet request. If you specified LaunchSpecifications in your Spot Fleet request, then omit this parameter.

onDemandTargetCapacity :: Maybe Int

The number of On-Demand Instances in the fleet.

context :: Maybe Text

Reserved.

excessCapacityTerminationPolicy :: Maybe ExcessCapacityTerminationPolicy

Indicates whether running Spot Instances should be terminated if the target capacity of the Spot Fleet request is decreased below the current size of the Spot Fleet.

targetCapacity :: Maybe Int

The size of the fleet.

spotFleetRequestId :: Text

The ID of the Spot Fleet request.

newModifySpotFleetRequest

Arguments

:: Text
$sel:spotFleetRequestId:ModifySpotFleetRequest'

-> ModifySpotFleetRequest

Create a value of ModifySpotFleetRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:launchTemplateConfigs:ModifySpotFleetRequest', modifySpotFleetRequest_launchTemplateConfigs - The launch template and overrides. You can only use this parameter if you specified a launch template (LaunchTemplateConfigs) in your Spot Fleet request. If you specified LaunchSpecifications in your Spot Fleet request, then omit this parameter.

$sel:onDemandTargetCapacity:ModifySpotFleetRequest', modifySpotFleetRequest_onDemandTargetCapacity - The number of On-Demand Instances in the fleet.

$sel:context:ModifySpotFleetRequest', modifySpotFleetRequest_context - Reserved.

$sel:excessCapacityTerminationPolicy:ModifySpotFleetRequest', modifySpotFleetRequest_excessCapacityTerminationPolicy - Indicates whether running Spot Instances should be terminated if the target capacity of the Spot Fleet request is decreased below the current size of the Spot Fleet.

$sel:targetCapacity:ModifySpotFleetRequest', modifySpotFleetRequest_targetCapacity - The size of the fleet.

$sel:spotFleetRequestId:ModifySpotFleetRequest', modifySpotFleetRequest_spotFleetRequestId - The ID of the Spot Fleet request.

Request Lenses

modifySpotFleetRequest_launchTemplateConfigs :: Lens' ModifySpotFleetRequest (Maybe [LaunchTemplateConfig])

The launch template and overrides. You can only use this parameter if you specified a launch template (LaunchTemplateConfigs) in your Spot Fleet request. If you specified LaunchSpecifications in your Spot Fleet request, then omit this parameter.

modifySpotFleetRequest_onDemandTargetCapacity :: Lens' ModifySpotFleetRequest (Maybe Int)

The number of On-Demand Instances in the fleet.

modifySpotFleetRequest_context :: Lens' ModifySpotFleetRequest (Maybe Text)

Reserved.

modifySpotFleetRequest_excessCapacityTerminationPolicy :: Lens' ModifySpotFleetRequest (Maybe ExcessCapacityTerminationPolicy)

Indicates whether running Spot Instances should be terminated if the target capacity of the Spot Fleet request is decreased below the current size of the Spot Fleet.

modifySpotFleetRequest_targetCapacity :: Lens' ModifySpotFleetRequest (Maybe Int)

The size of the fleet.

modifySpotFleetRequest_spotFleetRequestId :: Lens' ModifySpotFleetRequest Text

The ID of the Spot Fleet request.

Destructuring the Response

data ModifySpotFleetRequestResponse

Contains the output of ModifySpotFleetRequest.

See: newModifySpotFleetRequestResponse smart constructor.

Constructors

ModifySpotFleetRequestResponse'

Fields

return' :: Maybe Bool

Is true if the request succeeds, and an error otherwise.

httpStatus :: Int

The response's http status code.

newModifySpotFleetRequestResponse

Arguments

:: Int
$sel:httpStatus:ModifySpotFleetRequestResponse'

-> ModifySpotFleetRequestResponse

Create a value of ModifySpotFleetRequestResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:return':ModifySpotFleetRequestResponse', modifySpotFleetRequestResponse_return - Is true if the request succeeds, and an error otherwise.

$sel:httpStatus:ModifySpotFleetRequestResponse', modifySpotFleetRequestResponse_httpStatus - The response's http status code.

Response Lenses

modifySpotFleetRequestResponse_return :: Lens' ModifySpotFleetRequestResponse (Maybe Bool)

Is true if the request succeeds, and an error otherwise.

modifySpotFleetRequestResponse_httpStatus :: Lens' ModifySpotFleetRequestResponse Int

The response's http status code.