Network.AWS.EC2.ModifySubnetAttribute

Description

Modifies a subnet attribute. You can only modify one attribute at a time.

Creating a Request

data ModifySubnetAttribute

See: newModifySubnetAttribute smart constructor.

Constructors

ModifySubnetAttribute'

Fields

customerOwnedIpv4Pool :: Maybe Text

The customer-owned IPv4 address pool associated with the subnet.

You must set this value when you specify true for MapCustomerOwnedIpOnLaunch.

assignIpv6AddressOnCreation :: Maybe AttributeBooleanValue

Specify true to indicate that network interfaces created in the specified subnet should be assigned an IPv6 address. This includes a network interface that's created when launching an instance into the subnet (the instance therefore receives an IPv6 address).

If you enable the IPv6 addressing feature for your subnet, your network interface or instance only receives an IPv6 address if it's created using version 2016-11-15 or later of the Amazon EC2 API.

mapPublicIpOnLaunch :: Maybe AttributeBooleanValue

Specify true to indicate that network interfaces attached to instances created in the specified subnet should be assigned a public IPv4 address.

mapCustomerOwnedIpOnLaunch :: Maybe AttributeBooleanValue

Specify true to indicate that network interfaces attached to instances created in the specified subnet should be assigned a customer-owned IPv4 address.

When this value is true, you must specify the customer-owned IP pool using CustomerOwnedIpv4Pool.

subnetId :: Text

The ID of the subnet.

newModifySubnetAttribute

Arguments

:: Text
$sel:subnetId:ModifySubnetAttribute'

-> ModifySubnetAttribute

Create a value of ModifySubnetAttribute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:customerOwnedIpv4Pool:ModifySubnetAttribute', modifySubnetAttribute_customerOwnedIpv4Pool - The customer-owned IPv4 address pool associated with the subnet.

You must set this value when you specify true for MapCustomerOwnedIpOnLaunch.

$sel:assignIpv6AddressOnCreation:ModifySubnetAttribute', modifySubnetAttribute_assignIpv6AddressOnCreation - Specify true to indicate that network interfaces created in the specified subnet should be assigned an IPv6 address. This includes a network interface that's created when launching an instance into the subnet (the instance therefore receives an IPv6 address).

If you enable the IPv6 addressing feature for your subnet, your network interface or instance only receives an IPv6 address if it's created using version 2016-11-15 or later of the Amazon EC2 API.

$sel:mapPublicIpOnLaunch:ModifySubnetAttribute', modifySubnetAttribute_mapPublicIpOnLaunch - Specify true to indicate that network interfaces attached to instances created in the specified subnet should be assigned a public IPv4 address.

$sel:mapCustomerOwnedIpOnLaunch:ModifySubnetAttribute', modifySubnetAttribute_mapCustomerOwnedIpOnLaunch - Specify true to indicate that network interfaces attached to instances created in the specified subnet should be assigned a customer-owned IPv4 address.

When this value is true, you must specify the customer-owned IP pool using CustomerOwnedIpv4Pool.

$sel:subnetId:ModifySubnetAttribute', modifySubnetAttribute_subnetId - The ID of the subnet.

Request Lenses

modifySubnetAttribute_customerOwnedIpv4Pool :: Lens' ModifySubnetAttribute (Maybe Text)

The customer-owned IPv4 address pool associated with the subnet.

You must set this value when you specify true for MapCustomerOwnedIpOnLaunch.

modifySubnetAttribute_assignIpv6AddressOnCreation :: Lens' ModifySubnetAttribute (Maybe AttributeBooleanValue)

Specify true to indicate that network interfaces created in the specified subnet should be assigned an IPv6 address. This includes a network interface that's created when launching an instance into the subnet (the instance therefore receives an IPv6 address).

If you enable the IPv6 addressing feature for your subnet, your network interface or instance only receives an IPv6 address if it's created using version 2016-11-15 or later of the Amazon EC2 API.

modifySubnetAttribute_mapPublicIpOnLaunch :: Lens' ModifySubnetAttribute (Maybe AttributeBooleanValue)

Specify true to indicate that network interfaces attached to instances created in the specified subnet should be assigned a public IPv4 address.

modifySubnetAttribute_mapCustomerOwnedIpOnLaunch :: Lens' ModifySubnetAttribute (Maybe AttributeBooleanValue)

Specify true to indicate that network interfaces attached to instances created in the specified subnet should be assigned a customer-owned IPv4 address.

When this value is true, you must specify the customer-owned IP pool using CustomerOwnedIpv4Pool.

modifySubnetAttribute_subnetId :: Lens' ModifySubnetAttribute Text

The ID of the subnet.

Destructuring the Response

data ModifySubnetAttributeResponse

See: newModifySubnetAttributeResponse smart constructor.

Constructors

ModifySubnetAttributeResponse'

newModifySubnetAttributeResponse :: ModifySubnetAttributeResponse

Create a value of ModifySubnetAttributeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.