Network.AWS.EC2.ModifyTrafficMirrorFilterNetworkServices

Description

Allows or restricts mirroring network services.

By default, Amazon DNS network services are not eligible for Traffic Mirror. Use AddNetworkServices to add network services to a Traffic Mirror filter. When a network service is added to the Traffic Mirror filter, all traffic related to that network service will be mirrored. When you no longer want to mirror network services, use RemoveNetworkServices to remove the network services from the Traffic Mirror filter.

Creating a Request

data ModifyTrafficMirrorFilterNetworkServices

See: newModifyTrafficMirrorFilterNetworkServices smart constructor.

Constructors

ModifyTrafficMirrorFilterNetworkServices'

Fields

addNetworkServices :: Maybe [TrafficMirrorNetworkService]

The network service, for example Amazon DNS, that you want to mirror.

removeNetworkServices :: Maybe [TrafficMirrorNetworkService]

The network service, for example Amazon DNS, that you no longer want to mirror.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

trafficMirrorFilterId :: Text

The ID of the Traffic Mirror filter.

newModifyTrafficMirrorFilterNetworkServices

Arguments

:: Text
$sel:trafficMirrorFilterId:ModifyTrafficMirrorFilterNetworkServices'

-> ModifyTrafficMirrorFilterNetworkServices

Create a value of ModifyTrafficMirrorFilterNetworkServices with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:addNetworkServices:ModifyTrafficMirrorFilterNetworkServices', modifyTrafficMirrorFilterNetworkServices_addNetworkServices - The network service, for example Amazon DNS, that you want to mirror.

$sel:removeNetworkServices:ModifyTrafficMirrorFilterNetworkServices', modifyTrafficMirrorFilterNetworkServices_removeNetworkServices - The network service, for example Amazon DNS, that you no longer want to mirror.

$sel:dryRun:ModifyTrafficMirrorFilterNetworkServices', modifyTrafficMirrorFilterNetworkServices_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:trafficMirrorFilterId:ModifyTrafficMirrorFilterNetworkServices', modifyTrafficMirrorFilterNetworkServices_trafficMirrorFilterId - The ID of the Traffic Mirror filter.

Request Lenses

modifyTrafficMirrorFilterNetworkServices_addNetworkServices :: Lens' ModifyTrafficMirrorFilterNetworkServices (Maybe [TrafficMirrorNetworkService])

The network service, for example Amazon DNS, that you want to mirror.

modifyTrafficMirrorFilterNetworkServices_removeNetworkServices :: Lens' ModifyTrafficMirrorFilterNetworkServices (Maybe [TrafficMirrorNetworkService])

The network service, for example Amazon DNS, that you no longer want to mirror.

modifyTrafficMirrorFilterNetworkServices_dryRun :: Lens' ModifyTrafficMirrorFilterNetworkServices (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

modifyTrafficMirrorFilterNetworkServices_trafficMirrorFilterId :: Lens' ModifyTrafficMirrorFilterNetworkServices Text

The ID of the Traffic Mirror filter.

Destructuring the Response

data ModifyTrafficMirrorFilterNetworkServicesResponse

See: newModifyTrafficMirrorFilterNetworkServicesResponse smart constructor.

Constructors

ModifyTrafficMirrorFilterNetworkServicesResponse'

Fields

trafficMirrorFilter :: Maybe TrafficMirrorFilter

The Traffic Mirror filter that the network service is associated with.

httpStatus :: Int

The response's http status code.

newModifyTrafficMirrorFilterNetworkServicesResponse

Arguments

:: Int
$sel:httpStatus:ModifyTrafficMirrorFilterNetworkServicesResponse'

-> ModifyTrafficMirrorFilterNetworkServicesResponse

Create a value of ModifyTrafficMirrorFilterNetworkServicesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:trafficMirrorFilter:ModifyTrafficMirrorFilterNetworkServicesResponse', modifyTrafficMirrorFilterNetworkServicesResponse_trafficMirrorFilter - The Traffic Mirror filter that the network service is associated with.

$sel:httpStatus:ModifyTrafficMirrorFilterNetworkServicesResponse', modifyTrafficMirrorFilterNetworkServicesResponse_httpStatus - The response's http status code.

Response Lenses

modifyTrafficMirrorFilterNetworkServicesResponse_trafficMirrorFilter :: Lens' ModifyTrafficMirrorFilterNetworkServicesResponse (Maybe TrafficMirrorFilter)

The Traffic Mirror filter that the network service is associated with.

modifyTrafficMirrorFilterNetworkServicesResponse_httpStatus :: Lens' ModifyTrafficMirrorFilterNetworkServicesResponse Int

The response's http status code.