Network.AWS.EC2.ModifyTrafficMirrorFilterRule

Description

Modifies the specified Traffic Mirror rule.

DestinationCidrBlock and SourceCidrBlock must both be an IPv4 range or an IPv6 range.

Creating a Request

data ModifyTrafficMirrorFilterRule

See: newModifyTrafficMirrorFilterRule smart constructor.

Constructors

ModifyTrafficMirrorFilterRule'

Fields

removeFields :: Maybe [TrafficMirrorFilterRuleField]

The properties that you want to remove from the Traffic Mirror filter rule.

When you remove a property from a Traffic Mirror filter rule, the property is set to the default.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

sourcePortRange :: Maybe TrafficMirrorPortRangeRequest

The port range to assign to the Traffic Mirror rule.

ruleAction :: Maybe TrafficMirrorRuleAction

The action to assign to the rule.

sourceCidrBlock :: Maybe Text

The source CIDR block to assign to the Traffic Mirror rule.

trafficDirection :: Maybe TrafficDirection

The type of traffic (ingress | egress) to assign to the rule.

destinationCidrBlock :: Maybe Text

The destination CIDR block to assign to the Traffic Mirror rule.

protocol :: Maybe Int

The protocol, for example TCP, to assign to the Traffic Mirror rule.

description :: Maybe Text

The description to assign to the Traffic Mirror rule.

ruleNumber :: Maybe Int

The number of the Traffic Mirror rule. This number must be unique for each Traffic Mirror rule in a given direction. The rules are processed in ascending order by rule number.

destinationPortRange :: Maybe TrafficMirrorPortRangeRequest

The destination ports that are associated with the Traffic Mirror rule.

trafficMirrorFilterRuleId :: Text

The ID of the Traffic Mirror rule.

newModifyTrafficMirrorFilterRule

Arguments

:: Text
$sel:trafficMirrorFilterRuleId:ModifyTrafficMirrorFilterRule'

-> ModifyTrafficMirrorFilterRule

Create a value of ModifyTrafficMirrorFilterRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:removeFields:ModifyTrafficMirrorFilterRule', modifyTrafficMirrorFilterRule_removeFields - The properties that you want to remove from the Traffic Mirror filter rule.

When you remove a property from a Traffic Mirror filter rule, the property is set to the default.

$sel:dryRun:ModifyTrafficMirrorFilterRule', modifyTrafficMirrorFilterRule_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:sourcePortRange:ModifyTrafficMirrorFilterRule', modifyTrafficMirrorFilterRule_sourcePortRange - The port range to assign to the Traffic Mirror rule.

$sel:ruleAction:ModifyTrafficMirrorFilterRule', modifyTrafficMirrorFilterRule_ruleAction - The action to assign to the rule.

$sel:sourceCidrBlock:ModifyTrafficMirrorFilterRule', modifyTrafficMirrorFilterRule_sourceCidrBlock - The source CIDR block to assign to the Traffic Mirror rule.

$sel:trafficDirection:ModifyTrafficMirrorFilterRule', modifyTrafficMirrorFilterRule_trafficDirection - The type of traffic (ingress | egress) to assign to the rule.

$sel:destinationCidrBlock:ModifyTrafficMirrorFilterRule', modifyTrafficMirrorFilterRule_destinationCidrBlock - The destination CIDR block to assign to the Traffic Mirror rule.

$sel:protocol:ModifyTrafficMirrorFilterRule', modifyTrafficMirrorFilterRule_protocol - The protocol, for example TCP, to assign to the Traffic Mirror rule.

$sel:description:ModifyTrafficMirrorFilterRule', modifyTrafficMirrorFilterRule_description - The description to assign to the Traffic Mirror rule.

$sel:ruleNumber:ModifyTrafficMirrorFilterRule', modifyTrafficMirrorFilterRule_ruleNumber - The number of the Traffic Mirror rule. This number must be unique for each Traffic Mirror rule in a given direction. The rules are processed in ascending order by rule number.

$sel:destinationPortRange:ModifyTrafficMirrorFilterRule', modifyTrafficMirrorFilterRule_destinationPortRange - The destination ports that are associated with the Traffic Mirror rule.

$sel:trafficMirrorFilterRuleId:ModifyTrafficMirrorFilterRule', modifyTrafficMirrorFilterRule_trafficMirrorFilterRuleId - The ID of the Traffic Mirror rule.

Request Lenses

modifyTrafficMirrorFilterRule_removeFields :: Lens' ModifyTrafficMirrorFilterRule (Maybe [TrafficMirrorFilterRuleField])

The properties that you want to remove from the Traffic Mirror filter rule.

When you remove a property from a Traffic Mirror filter rule, the property is set to the default.

modifyTrafficMirrorFilterRule_dryRun :: Lens' ModifyTrafficMirrorFilterRule (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

modifyTrafficMirrorFilterRule_sourcePortRange :: Lens' ModifyTrafficMirrorFilterRule (Maybe TrafficMirrorPortRangeRequest)

The port range to assign to the Traffic Mirror rule.

modifyTrafficMirrorFilterRule_ruleAction :: Lens' ModifyTrafficMirrorFilterRule (Maybe TrafficMirrorRuleAction)

The action to assign to the rule.

modifyTrafficMirrorFilterRule_sourceCidrBlock :: Lens' ModifyTrafficMirrorFilterRule (Maybe Text)

The source CIDR block to assign to the Traffic Mirror rule.

modifyTrafficMirrorFilterRule_trafficDirection :: Lens' ModifyTrafficMirrorFilterRule (Maybe TrafficDirection)

The type of traffic (ingress | egress) to assign to the rule.

modifyTrafficMirrorFilterRule_destinationCidrBlock :: Lens' ModifyTrafficMirrorFilterRule (Maybe Text)

The destination CIDR block to assign to the Traffic Mirror rule.

modifyTrafficMirrorFilterRule_protocol :: Lens' ModifyTrafficMirrorFilterRule (Maybe Int)

The protocol, for example TCP, to assign to the Traffic Mirror rule.

modifyTrafficMirrorFilterRule_description :: Lens' ModifyTrafficMirrorFilterRule (Maybe Text)

The description to assign to the Traffic Mirror rule.

modifyTrafficMirrorFilterRule_ruleNumber :: Lens' ModifyTrafficMirrorFilterRule (Maybe Int)

The number of the Traffic Mirror rule. This number must be unique for each Traffic Mirror rule in a given direction. The rules are processed in ascending order by rule number.

modifyTrafficMirrorFilterRule_destinationPortRange :: Lens' ModifyTrafficMirrorFilterRule (Maybe TrafficMirrorPortRangeRequest)

The destination ports that are associated with the Traffic Mirror rule.

modifyTrafficMirrorFilterRule_trafficMirrorFilterRuleId :: Lens' ModifyTrafficMirrorFilterRule Text

The ID of the Traffic Mirror rule.

Destructuring the Response

data ModifyTrafficMirrorFilterRuleResponse

See: newModifyTrafficMirrorFilterRuleResponse smart constructor.

Constructors

ModifyTrafficMirrorFilterRuleResponse'

Fields

trafficMirrorFilterRule :: Maybe TrafficMirrorFilterRule

Modifies a Traffic Mirror rule.

httpStatus :: Int

The response's http status code.

newModifyTrafficMirrorFilterRuleResponse

Arguments

:: Int
$sel:httpStatus:ModifyTrafficMirrorFilterRuleResponse'

-> ModifyTrafficMirrorFilterRuleResponse

Create a value of ModifyTrafficMirrorFilterRuleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:trafficMirrorFilterRule:ModifyTrafficMirrorFilterRuleResponse', modifyTrafficMirrorFilterRuleResponse_trafficMirrorFilterRule - Modifies a Traffic Mirror rule.

$sel:httpStatus:ModifyTrafficMirrorFilterRuleResponse', modifyTrafficMirrorFilterRuleResponse_httpStatus - The response's http status code.

Response Lenses

modifyTrafficMirrorFilterRuleResponse_trafficMirrorFilterRule :: Lens' ModifyTrafficMirrorFilterRuleResponse (Maybe TrafficMirrorFilterRule)

Modifies a Traffic Mirror rule.

modifyTrafficMirrorFilterRuleResponse_httpStatus :: Lens' ModifyTrafficMirrorFilterRuleResponse Int

The response's http status code.