Network.AWS.EC2.ModifyVpcEndpoint

Description

Modifies attributes of a specified VPC endpoint. The attributes that you can modify depend on the type of VPC endpoint (interface, gateway, or Gateway Load Balancer). For more information, see VPC Endpoints in the Amazon Virtual Private Cloud User Guide.

Creating a Request

data ModifyVpcEndpoint

Contains the parameters for ModifyVpcEndpoint.

See: newModifyVpcEndpoint smart constructor.

Constructors

ModifyVpcEndpoint'

Fields

policyDocument :: Maybe Text

(Interface and gateway endpoints) A policy to attach to the endpoint that controls access to the service. The policy must be in valid JSON format.

removeRouteTableIds :: Maybe [Text]

(Gateway endpoint) One or more route table IDs to disassociate from the endpoint.

resetPolicy :: Maybe Bool

(Gateway endpoint) Specify true to reset the policy document to the default policy. The default policy allows full access to the service.

addRouteTableIds :: Maybe [Text]

(Gateway endpoint) One or more route tables IDs to associate with the endpoint.

privateDnsEnabled :: Maybe Bool

(Interface endpoint) Indicates whether a private hosted zone is associated with the VPC.

addSubnetIds :: Maybe [Text]

(Interface and Gateway Load Balancer endpoints) One or more subnet IDs in which to serve the endpoint. For a Gateway Load Balancer endpoint, you can specify only one subnet.

removeSubnetIds :: Maybe [Text]

(Interface endpoint) One or more subnets IDs in which to remove the endpoint.

addSecurityGroupIds :: Maybe [Text]

(Interface endpoint) One or more security group IDs to associate with the network interface.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

removeSecurityGroupIds :: Maybe [Text]

(Interface endpoint) One or more security group IDs to disassociate from the network interface.

vpcEndpointId :: Text

The ID of the endpoint.

newModifyVpcEndpoint

Arguments

:: Text
$sel:vpcEndpointId:ModifyVpcEndpoint'

-> ModifyVpcEndpoint

Create a value of ModifyVpcEndpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policyDocument:ModifyVpcEndpoint', modifyVpcEndpoint_policyDocument - (Interface and gateway endpoints) A policy to attach to the endpoint that controls access to the service. The policy must be in valid JSON format.

$sel:removeRouteTableIds:ModifyVpcEndpoint', modifyVpcEndpoint_removeRouteTableIds - (Gateway endpoint) One or more route table IDs to disassociate from the endpoint.

$sel:resetPolicy:ModifyVpcEndpoint', modifyVpcEndpoint_resetPolicy - (Gateway endpoint) Specify true to reset the policy document to the default policy. The default policy allows full access to the service.

$sel:addRouteTableIds:ModifyVpcEndpoint', modifyVpcEndpoint_addRouteTableIds - (Gateway endpoint) One or more route tables IDs to associate with the endpoint.

$sel:privateDnsEnabled:ModifyVpcEndpoint', modifyVpcEndpoint_privateDnsEnabled - (Interface endpoint) Indicates whether a private hosted zone is associated with the VPC.

$sel:addSubnetIds:ModifyVpcEndpoint', modifyVpcEndpoint_addSubnetIds - (Interface and Gateway Load Balancer endpoints) One or more subnet IDs in which to serve the endpoint. For a Gateway Load Balancer endpoint, you can specify only one subnet.

$sel:removeSubnetIds:ModifyVpcEndpoint', modifyVpcEndpoint_removeSubnetIds - (Interface endpoint) One or more subnets IDs in which to remove the endpoint.

$sel:addSecurityGroupIds:ModifyVpcEndpoint', modifyVpcEndpoint_addSecurityGroupIds - (Interface endpoint) One or more security group IDs to associate with the network interface.

$sel:dryRun:ModifyVpcEndpoint', modifyVpcEndpoint_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:removeSecurityGroupIds:ModifyVpcEndpoint', modifyVpcEndpoint_removeSecurityGroupIds - (Interface endpoint) One or more security group IDs to disassociate from the network interface.

$sel:vpcEndpointId:ModifyVpcEndpoint', modifyVpcEndpoint_vpcEndpointId - The ID of the endpoint.

Request Lenses

modifyVpcEndpoint_policyDocument :: Lens' ModifyVpcEndpoint (Maybe Text)

(Interface and gateway endpoints) A policy to attach to the endpoint that controls access to the service. The policy must be in valid JSON format.

modifyVpcEndpoint_removeRouteTableIds :: Lens' ModifyVpcEndpoint (Maybe [Text])

(Gateway endpoint) One or more route table IDs to disassociate from the endpoint.

modifyVpcEndpoint_resetPolicy :: Lens' ModifyVpcEndpoint (Maybe Bool)

(Gateway endpoint) Specify true to reset the policy document to the default policy. The default policy allows full access to the service.

modifyVpcEndpoint_addRouteTableIds :: Lens' ModifyVpcEndpoint (Maybe [Text])

(Gateway endpoint) One or more route tables IDs to associate with the endpoint.

modifyVpcEndpoint_privateDnsEnabled :: Lens' ModifyVpcEndpoint (Maybe Bool)

(Interface endpoint) Indicates whether a private hosted zone is associated with the VPC.

modifyVpcEndpoint_addSubnetIds :: Lens' ModifyVpcEndpoint (Maybe [Text])

(Interface and Gateway Load Balancer endpoints) One or more subnet IDs in which to serve the endpoint. For a Gateway Load Balancer endpoint, you can specify only one subnet.

modifyVpcEndpoint_removeSubnetIds :: Lens' ModifyVpcEndpoint (Maybe [Text])

(Interface endpoint) One or more subnets IDs in which to remove the endpoint.

modifyVpcEndpoint_addSecurityGroupIds :: Lens' ModifyVpcEndpoint (Maybe [Text])

(Interface endpoint) One or more security group IDs to associate with the network interface.

modifyVpcEndpoint_dryRun :: Lens' ModifyVpcEndpoint (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

modifyVpcEndpoint_removeSecurityGroupIds :: Lens' ModifyVpcEndpoint (Maybe [Text])

(Interface endpoint) One or more security group IDs to disassociate from the network interface.

modifyVpcEndpoint_vpcEndpointId :: Lens' ModifyVpcEndpoint Text

The ID of the endpoint.

Destructuring the Response

data ModifyVpcEndpointResponse

See: newModifyVpcEndpointResponse smart constructor.

Constructors

ModifyVpcEndpointResponse'

Fields

return' :: Maybe Bool

Returns true if the request succeeds; otherwise, it returns an error.

httpStatus :: Int

The response's http status code.

newModifyVpcEndpointResponse

Arguments

:: Int
$sel:httpStatus:ModifyVpcEndpointResponse'

-> ModifyVpcEndpointResponse

Create a value of ModifyVpcEndpointResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:return':ModifyVpcEndpointResponse', modifyVpcEndpointResponse_return - Returns true if the request succeeds; otherwise, it returns an error.

$sel:httpStatus:ModifyVpcEndpointResponse', modifyVpcEndpointResponse_httpStatus - The response's http status code.

Response Lenses

modifyVpcEndpointResponse_return :: Lens' ModifyVpcEndpointResponse (Maybe Bool)

Returns true if the request succeeds; otherwise, it returns an error.

modifyVpcEndpointResponse_httpStatus :: Lens' ModifyVpcEndpointResponse Int

The response's http status code.