Network.AWS.EC2.ModifyVpcEndpointServiceConfiguration

Description

Modifies the attributes of your VPC endpoint service configuration. You can change the Network Load Balancers or Gateway Load Balancers for your service, and you can specify whether acceptance is required for requests to connect to your endpoint service through an interface VPC endpoint.

If you set or modify the private DNS name, you must prove that you own the private DNS domain name. For more information, see VPC Endpoint Service Private DNS Name Verification in the Amazon Virtual Private Cloud User Guide.

Creating a Request

data ModifyVpcEndpointServiceConfiguration

See: newModifyVpcEndpointServiceConfiguration smart constructor.

Constructors

ModifyVpcEndpointServiceConfiguration'

Fields

removeGatewayLoadBalancerArns :: Maybe [Text]

The Amazon Resource Names (ARNs) of Gateway Load Balancers to remove from your service configuration.

removePrivateDnsName :: Maybe Bool

(Interface endpoint configuration) Removes the private DNS name of the endpoint service.

addGatewayLoadBalancerArns :: Maybe [Text]

The Amazon Resource Names (ARNs) of Gateway Load Balancers to add to your service configuration.

removeNetworkLoadBalancerArns :: Maybe [Text]

The Amazon Resource Names (ARNs) of Network Load Balancers to remove from your service configuration.

acceptanceRequired :: Maybe Bool

Indicates whether requests to create an endpoint to your service must be accepted.

addNetworkLoadBalancerArns :: Maybe [Text]

The Amazon Resource Names (ARNs) of Network Load Balancers to add to your service configuration.

privateDnsName :: Maybe Text

(Interface endpoint configuration) The private DNS name to assign to the endpoint service.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

serviceId :: Text

The ID of the service.

newModifyVpcEndpointServiceConfiguration

Arguments

:: Text
$sel:serviceId:ModifyVpcEndpointServiceConfiguration'

-> ModifyVpcEndpointServiceConfiguration

Create a value of ModifyVpcEndpointServiceConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:removeGatewayLoadBalancerArns:ModifyVpcEndpointServiceConfiguration', modifyVpcEndpointServiceConfiguration_removeGatewayLoadBalancerArns - The Amazon Resource Names (ARNs) of Gateway Load Balancers to remove from your service configuration.

$sel:removePrivateDnsName:ModifyVpcEndpointServiceConfiguration', modifyVpcEndpointServiceConfiguration_removePrivateDnsName - (Interface endpoint configuration) Removes the private DNS name of the endpoint service.

$sel:addGatewayLoadBalancerArns:ModifyVpcEndpointServiceConfiguration', modifyVpcEndpointServiceConfiguration_addGatewayLoadBalancerArns - The Amazon Resource Names (ARNs) of Gateway Load Balancers to add to your service configuration.

$sel:removeNetworkLoadBalancerArns:ModifyVpcEndpointServiceConfiguration', modifyVpcEndpointServiceConfiguration_removeNetworkLoadBalancerArns - The Amazon Resource Names (ARNs) of Network Load Balancers to remove from your service configuration.

$sel:acceptanceRequired:ModifyVpcEndpointServiceConfiguration', modifyVpcEndpointServiceConfiguration_acceptanceRequired - Indicates whether requests to create an endpoint to your service must be accepted.

$sel:addNetworkLoadBalancerArns:ModifyVpcEndpointServiceConfiguration', modifyVpcEndpointServiceConfiguration_addNetworkLoadBalancerArns - The Amazon Resource Names (ARNs) of Network Load Balancers to add to your service configuration.

$sel:privateDnsName:ModifyVpcEndpointServiceConfiguration', modifyVpcEndpointServiceConfiguration_privateDnsName - (Interface endpoint configuration) The private DNS name to assign to the endpoint service.

$sel:dryRun:ModifyVpcEndpointServiceConfiguration', modifyVpcEndpointServiceConfiguration_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:serviceId:ModifyVpcEndpointServiceConfiguration', modifyVpcEndpointServiceConfiguration_serviceId - The ID of the service.

Request Lenses

modifyVpcEndpointServiceConfiguration_removeGatewayLoadBalancerArns :: Lens' ModifyVpcEndpointServiceConfiguration (Maybe [Text])

The Amazon Resource Names (ARNs) of Gateway Load Balancers to remove from your service configuration.

modifyVpcEndpointServiceConfiguration_removePrivateDnsName :: Lens' ModifyVpcEndpointServiceConfiguration (Maybe Bool)

(Interface endpoint configuration) Removes the private DNS name of the endpoint service.

modifyVpcEndpointServiceConfiguration_addGatewayLoadBalancerArns :: Lens' ModifyVpcEndpointServiceConfiguration (Maybe [Text])

The Amazon Resource Names (ARNs) of Gateway Load Balancers to add to your service configuration.

modifyVpcEndpointServiceConfiguration_removeNetworkLoadBalancerArns :: Lens' ModifyVpcEndpointServiceConfiguration (Maybe [Text])

The Amazon Resource Names (ARNs) of Network Load Balancers to remove from your service configuration.

modifyVpcEndpointServiceConfiguration_acceptanceRequired :: Lens' ModifyVpcEndpointServiceConfiguration (Maybe Bool)

Indicates whether requests to create an endpoint to your service must be accepted.

modifyVpcEndpointServiceConfiguration_addNetworkLoadBalancerArns :: Lens' ModifyVpcEndpointServiceConfiguration (Maybe [Text])

The Amazon Resource Names (ARNs) of Network Load Balancers to add to your service configuration.

modifyVpcEndpointServiceConfiguration_privateDnsName :: Lens' ModifyVpcEndpointServiceConfiguration (Maybe Text)

(Interface endpoint configuration) The private DNS name to assign to the endpoint service.

modifyVpcEndpointServiceConfiguration_dryRun :: Lens' ModifyVpcEndpointServiceConfiguration (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

modifyVpcEndpointServiceConfiguration_serviceId :: Lens' ModifyVpcEndpointServiceConfiguration Text

The ID of the service.

Destructuring the Response

data ModifyVpcEndpointServiceConfigurationResponse

See: newModifyVpcEndpointServiceConfigurationResponse smart constructor.

Constructors

ModifyVpcEndpointServiceConfigurationResponse'

Fields

return' :: Maybe Bool

Returns true if the request succeeds; otherwise, it returns an error.

httpStatus :: Int

The response's http status code.

newModifyVpcEndpointServiceConfigurationResponse

Arguments

:: Int
$sel:httpStatus:ModifyVpcEndpointServiceConfigurationResponse'

-> ModifyVpcEndpointServiceConfigurationResponse

Create a value of ModifyVpcEndpointServiceConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:return':ModifyVpcEndpointServiceConfigurationResponse', modifyVpcEndpointServiceConfigurationResponse_return - Returns true if the request succeeds; otherwise, it returns an error.

$sel:httpStatus:ModifyVpcEndpointServiceConfigurationResponse', modifyVpcEndpointServiceConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

modifyVpcEndpointServiceConfigurationResponse_return :: Lens' ModifyVpcEndpointServiceConfigurationResponse (Maybe Bool)

Returns true if the request succeeds; otherwise, it returns an error.

modifyVpcEndpointServiceConfigurationResponse_httpStatus :: Lens' ModifyVpcEndpointServiceConfigurationResponse Int

The response's http status code.