Network.AWS.EC2.ModifyVpcPeeringConnectionOptions

Description

Modifies the VPC peering connection options on one side of a VPC peering connection. You can do the following:

Enable/disable communication over the peering connection between an EC2-Classic instance that's linked to your VPC (using ClassicLink) and instances in the peer VPC.

Enable/disable communication over the peering connection between instances in your VPC and an EC2-Classic instance that's linked to the peer VPC.

Enable/disable the ability to resolve public DNS hostnames to private IP addresses when queried from instances in the peer VPC.

If the peered VPCs are in the same Amazon Web Services account, you can enable DNS resolution for queries from the local VPC. This ensures that queries from the local VPC resolve to private IP addresses in the peer VPC. This option is not available if the peered VPCs are in different different Amazon Web Services accounts or different Regions. For peered VPCs in different Amazon Web Services accounts, each Amazon Web Services account owner must initiate a separate request to modify the peering connection options. For inter-region peering connections, you must use the Region for the requester VPC to modify the requester VPC peering options and the Region for the accepter VPC to modify the accepter VPC peering options. To verify which VPCs are the accepter and the requester for a VPC peering connection, use the DescribeVpcPeeringConnections command.

Creating a Request

data ModifyVpcPeeringConnectionOptions

See: newModifyVpcPeeringConnectionOptions smart constructor.

Constructors

ModifyVpcPeeringConnectionOptions'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

accepterPeeringConnectionOptions :: Maybe PeeringConnectionOptionsRequest

The VPC peering connection options for the accepter VPC.

requesterPeeringConnectionOptions :: Maybe PeeringConnectionOptionsRequest

The VPC peering connection options for the requester VPC.

vpcPeeringConnectionId :: Text

The ID of the VPC peering connection.

newModifyVpcPeeringConnectionOptions

Arguments

:: Text
$sel:vpcPeeringConnectionId:ModifyVpcPeeringConnectionOptions'

-> ModifyVpcPeeringConnectionOptions

Create a value of ModifyVpcPeeringConnectionOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:ModifyVpcPeeringConnectionOptions', modifyVpcPeeringConnectionOptions_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:accepterPeeringConnectionOptions:ModifyVpcPeeringConnectionOptions', modifyVpcPeeringConnectionOptions_accepterPeeringConnectionOptions - The VPC peering connection options for the accepter VPC.

$sel:requesterPeeringConnectionOptions:ModifyVpcPeeringConnectionOptions', modifyVpcPeeringConnectionOptions_requesterPeeringConnectionOptions - The VPC peering connection options for the requester VPC.

$sel:vpcPeeringConnectionId:ModifyVpcPeeringConnectionOptions', modifyVpcPeeringConnectionOptions_vpcPeeringConnectionId - The ID of the VPC peering connection.

Request Lenses

modifyVpcPeeringConnectionOptions_dryRun :: Lens' ModifyVpcPeeringConnectionOptions (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

modifyVpcPeeringConnectionOptions_accepterPeeringConnectionOptions :: Lens' ModifyVpcPeeringConnectionOptions (Maybe PeeringConnectionOptionsRequest)

The VPC peering connection options for the accepter VPC.

modifyVpcPeeringConnectionOptions_requesterPeeringConnectionOptions :: Lens' ModifyVpcPeeringConnectionOptions (Maybe PeeringConnectionOptionsRequest)

The VPC peering connection options for the requester VPC.

modifyVpcPeeringConnectionOptions_vpcPeeringConnectionId :: Lens' ModifyVpcPeeringConnectionOptions Text

The ID of the VPC peering connection.

Destructuring the Response

data ModifyVpcPeeringConnectionOptionsResponse

See: newModifyVpcPeeringConnectionOptionsResponse smart constructor.

Constructors

ModifyVpcPeeringConnectionOptionsResponse'

Fields

accepterPeeringConnectionOptions :: Maybe PeeringConnectionOptions

Information about the VPC peering connection options for the accepter VPC.

requesterPeeringConnectionOptions :: Maybe PeeringConnectionOptions

Information about the VPC peering connection options for the requester VPC.

httpStatus :: Int

The response's http status code.

newModifyVpcPeeringConnectionOptionsResponse

Arguments

:: Int
$sel:httpStatus:ModifyVpcPeeringConnectionOptionsResponse'

-> ModifyVpcPeeringConnectionOptionsResponse

Create a value of ModifyVpcPeeringConnectionOptionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accepterPeeringConnectionOptions:ModifyVpcPeeringConnectionOptions', modifyVpcPeeringConnectionOptionsResponse_accepterPeeringConnectionOptions - Information about the VPC peering connection options for the accepter VPC.

$sel:requesterPeeringConnectionOptions:ModifyVpcPeeringConnectionOptions', modifyVpcPeeringConnectionOptionsResponse_requesterPeeringConnectionOptions - Information about the VPC peering connection options for the requester VPC.

$sel:httpStatus:ModifyVpcPeeringConnectionOptionsResponse', modifyVpcPeeringConnectionOptionsResponse_httpStatus - The response's http status code.

Response Lenses

modifyVpcPeeringConnectionOptionsResponse_accepterPeeringConnectionOptions :: Lens' ModifyVpcPeeringConnectionOptionsResponse (Maybe PeeringConnectionOptions)

Information about the VPC peering connection options for the accepter VPC.

modifyVpcPeeringConnectionOptionsResponse_requesterPeeringConnectionOptions :: Lens' ModifyVpcPeeringConnectionOptionsResponse (Maybe PeeringConnectionOptions)

Information about the VPC peering connection options for the requester VPC.

modifyVpcPeeringConnectionOptionsResponse_httpStatus :: Lens' ModifyVpcPeeringConnectionOptionsResponse Int

The response's http status code.