Network.AWS.EC2.ModifyVpnConnection

Description

Modifies the customer gateway or the target gateway of an Amazon Web Services Site-to-Site VPN connection. To modify the target gateway, the following migration options are available:

An existing virtual private gateway to a new virtual private gateway

An existing virtual private gateway to a transit gateway

An existing transit gateway to a new transit gateway

An existing transit gateway to a virtual private gateway

Before you perform the migration to the new gateway, you must configure the new gateway. Use CreateVpnGateway to create a virtual private gateway, or CreateTransitGateway to create a transit gateway.

This step is required when you migrate from a virtual private gateway with static routes to a transit gateway.

You must delete the static routes before you migrate to the new gateway.

Keep a copy of the static route before you delete it. You will need to add back these routes to the transit gateway after the VPN connection migration is complete.

After you migrate to the new gateway, you might need to modify your VPC route table. Use CreateRoute and DeleteRoute to make the changes described in Update VPC route tables in the Amazon Web Services Site-to-Site VPN User Guide.

When the new gateway is a transit gateway, modify the transit gateway route table to allow traffic between the VPC and the Amazon Web Services Site-to-Site VPN connection. Use CreateTransitGatewayRoute to add the routes.

If you deleted VPN static routes, you must add the static routes to the transit gateway route table.

After you perform this operation, the VPN endpoint's IP addresses on the Amazon Web Services side and the tunnel options remain intact. Your Amazon Web Services Site-to-Site VPN connection will be temporarily unavailable for a brief period while we provision the new endpoints.

Creating a Request

data ModifyVpnConnection

See: newModifyVpnConnection smart constructor.

Constructors

ModifyVpnConnection'

Fields

vpnGatewayId :: Maybe Text

The ID of the virtual private gateway at the Amazon Web Services side of the VPN connection.

customerGatewayId :: Maybe Text

The ID of the customer gateway at your end of the VPN connection.

transitGatewayId :: Maybe Text

The ID of the transit gateway.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

vpnConnectionId :: Text

The ID of the VPN connection.

newModifyVpnConnection

Arguments

:: Text
$sel:vpnConnectionId:ModifyVpnConnection'

-> ModifyVpnConnection

Create a value of ModifyVpnConnection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpnGatewayId:ModifyVpnConnection', modifyVpnConnection_vpnGatewayId - The ID of the virtual private gateway at the Amazon Web Services side of the VPN connection.

$sel:customerGatewayId:ModifyVpnConnection', modifyVpnConnection_customerGatewayId - The ID of the customer gateway at your end of the VPN connection.

$sel:transitGatewayId:ModifyVpnConnection', modifyVpnConnection_transitGatewayId - The ID of the transit gateway.

$sel:dryRun:ModifyVpnConnection', modifyVpnConnection_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:vpnConnectionId:ModifyVpnConnection', modifyVpnConnection_vpnConnectionId - The ID of the VPN connection.

Request Lenses

modifyVpnConnection_vpnGatewayId :: Lens' ModifyVpnConnection (Maybe Text)

The ID of the virtual private gateway at the Amazon Web Services side of the VPN connection.

modifyVpnConnection_customerGatewayId :: Lens' ModifyVpnConnection (Maybe Text)

The ID of the customer gateway at your end of the VPN connection.

modifyVpnConnection_transitGatewayId :: Lens' ModifyVpnConnection (Maybe Text)

The ID of the transit gateway.

modifyVpnConnection_dryRun :: Lens' ModifyVpnConnection (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

modifyVpnConnection_vpnConnectionId :: Lens' ModifyVpnConnection Text

The ID of the VPN connection.

Destructuring the Response

data ModifyVpnConnectionResponse

See: newModifyVpnConnectionResponse smart constructor.

Constructors

ModifyVpnConnectionResponse'

Fields

vpnConnection :: Maybe VpnConnection

httpStatus :: Int

The response's http status code.

newModifyVpnConnectionResponse

Arguments

:: Int
$sel:httpStatus:ModifyVpnConnectionResponse'

-> ModifyVpnConnectionResponse

Create a value of ModifyVpnConnectionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpnConnection:ModifyVpnConnectionResponse', modifyVpnConnectionResponse_vpnConnection - Undocumented member.

$sel:httpStatus:ModifyVpnConnectionResponse', modifyVpnConnectionResponse_httpStatus - The response's http status code.

Response Lenses

modifyVpnConnectionResponse_vpnConnection :: Lens' ModifyVpnConnectionResponse (Maybe VpnConnection)

Undocumented member.

modifyVpnConnectionResponse_httpStatus :: Lens' ModifyVpnConnectionResponse Int

The response's http status code.