Network.AWS.EC2.ProvisionByoipCidr

Description

Provisions an IPv4 or IPv6 address range for use with your Amazon Web Services resources through bring your own IP addresses (BYOIP) and creates a corresponding address pool. After the address range is provisioned, it is ready to be advertised using AdvertiseByoipCidr.

Amazon Web Services verifies that you own the address range and are authorized to advertise it. You must ensure that the address range is registered to you and that you created an RPKI ROA to authorize Amazon ASNs 16509 and 14618 to advertise the address range. For more information, see Bring your own IP addresses (BYOIP) in the Amazon Elastic Compute Cloud User Guide.

Provisioning an address range is an asynchronous operation, so the call returns immediately, but the address range is not ready to use until its status changes from pending-provision to provisioned. To monitor the status of an address range, use DescribeByoipCidrs. To allocate an Elastic IP address from your IPv4 address pool, use AllocateAddress with either the specific address from the address pool or the ID of the address pool.

Creating a Request

data ProvisionByoipCidr

See: newProvisionByoipCidr smart constructor.

Constructors

ProvisionByoipCidr'

Fields

cidrAuthorizationContext :: Maybe CidrAuthorizationContext

A signed document that proves that you are authorized to bring the specified IP address range to Amazon using BYOIP.

poolTagSpecifications :: Maybe [TagSpecification]

The tags to apply to the address pool.

publiclyAdvertisable :: Maybe Bool

(IPv6 only) Indicate whether the address range will be publicly advertised to the internet.

Default: true

description :: Maybe Text

A description for the address range and the address pool.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

multiRegion :: Maybe Bool

Reserved.

cidr :: Text

The public IPv4 or IPv6 address range, in CIDR notation. The most specific IPv4 prefix that you can specify is /24. The most specific IPv6 prefix you can specify is /56. The address range cannot overlap with another address range that you've brought to this or another Region.

newProvisionByoipCidr

Arguments

:: Text
$sel:cidr:ProvisionByoipCidr'

-> ProvisionByoipCidr

Create a value of ProvisionByoipCidr with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cidrAuthorizationContext:ProvisionByoipCidr', provisionByoipCidr_cidrAuthorizationContext - A signed document that proves that you are authorized to bring the specified IP address range to Amazon using BYOIP.

$sel:poolTagSpecifications:ProvisionByoipCidr', provisionByoipCidr_poolTagSpecifications - The tags to apply to the address pool.

$sel:publiclyAdvertisable:ProvisionByoipCidr', provisionByoipCidr_publiclyAdvertisable - (IPv6 only) Indicate whether the address range will be publicly advertised to the internet.

Default: true

$sel:description:ProvisionByoipCidr', provisionByoipCidr_description - A description for the address range and the address pool.

$sel:dryRun:ProvisionByoipCidr', provisionByoipCidr_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:multiRegion:ProvisionByoipCidr', provisionByoipCidr_multiRegion - Reserved.

$sel:cidr:ProvisionByoipCidr', provisionByoipCidr_cidr - The public IPv4 or IPv6 address range, in CIDR notation. The most specific IPv4 prefix that you can specify is /24. The most specific IPv6 prefix you can specify is /56. The address range cannot overlap with another address range that you've brought to this or another Region.

Request Lenses

provisionByoipCidr_cidrAuthorizationContext :: Lens' ProvisionByoipCidr (Maybe CidrAuthorizationContext)

A signed document that proves that you are authorized to bring the specified IP address range to Amazon using BYOIP.

provisionByoipCidr_poolTagSpecifications :: Lens' ProvisionByoipCidr (Maybe [TagSpecification])

The tags to apply to the address pool.

provisionByoipCidr_publiclyAdvertisable :: Lens' ProvisionByoipCidr (Maybe Bool)

(IPv6 only) Indicate whether the address range will be publicly advertised to the internet.

Default: true

provisionByoipCidr_description :: Lens' ProvisionByoipCidr (Maybe Text)

A description for the address range and the address pool.

provisionByoipCidr_dryRun :: Lens' ProvisionByoipCidr (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

provisionByoipCidr_multiRegion :: Lens' ProvisionByoipCidr (Maybe Bool)

Reserved.

provisionByoipCidr_cidr :: Lens' ProvisionByoipCidr Text

The public IPv4 or IPv6 address range, in CIDR notation. The most specific IPv4 prefix that you can specify is /24. The most specific IPv6 prefix you can specify is /56. The address range cannot overlap with another address range that you've brought to this or another Region.

Destructuring the Response

data ProvisionByoipCidrResponse

See: newProvisionByoipCidrResponse smart constructor.

Constructors

ProvisionByoipCidrResponse'

Fields

byoipCidr :: Maybe ByoipCidr

Information about the address range.

httpStatus :: Int

The response's http status code.

newProvisionByoipCidrResponse

Arguments

:: Int
$sel:httpStatus:ProvisionByoipCidrResponse'

-> ProvisionByoipCidrResponse

Create a value of ProvisionByoipCidrResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:byoipCidr:ProvisionByoipCidrResponse', provisionByoipCidrResponse_byoipCidr - Information about the address range.

$sel:httpStatus:ProvisionByoipCidrResponse', provisionByoipCidrResponse_httpStatus - The response's http status code.

Response Lenses

provisionByoipCidrResponse_byoipCidr :: Lens' ProvisionByoipCidrResponse (Maybe ByoipCidr)

Information about the address range.

provisionByoipCidrResponse_httpStatus :: Lens' ProvisionByoipCidrResponse Int

The response's http status code.