Network.AWS.EC2.PurchaseReservedInstancesOffering

Description

Purchases a Reserved Instance for use with your account. With Reserved Instances, you pay a lower hourly rate compared to On-Demand instance pricing.

Use DescribeReservedInstancesOfferings to get a list of Reserved Instance offerings that match your specifications. After you've purchased a Reserved Instance, you can check for your new Reserved Instance with DescribeReservedInstances.

To queue a purchase for a future date and time, specify a purchase time. If you do not specify a purchase time, the default is the current time.

For more information, see Reserved Instances and Reserved Instance Marketplace in the Amazon EC2 User Guide.

Creating a Request

data PurchaseReservedInstancesOffering

Contains the parameters for PurchaseReservedInstancesOffering.

See: newPurchaseReservedInstancesOffering smart constructor.

Constructors

PurchaseReservedInstancesOffering'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

purchaseTime :: Maybe ISO8601

The time at which to purchase the Reserved Instance, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ).

limitPrice :: Maybe ReservedInstanceLimitPrice

Specified for Reserved Instance Marketplace offerings to limit the total order and ensure that the Reserved Instances are not purchased at unexpected prices.

instanceCount :: Int

The number of Reserved Instances to purchase.

reservedInstancesOfferingId :: Text

The ID of the Reserved Instance offering to purchase.

newPurchaseReservedInstancesOffering

Arguments

:: Int
$sel:instanceCount:PurchaseReservedInstancesOffering'

-> Text
$sel:reservedInstancesOfferingId:PurchaseReservedInstancesOffering'

-> PurchaseReservedInstancesOffering

Create a value of PurchaseReservedInstancesOffering with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:PurchaseReservedInstancesOffering', purchaseReservedInstancesOffering_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:purchaseTime:PurchaseReservedInstancesOffering', purchaseReservedInstancesOffering_purchaseTime - The time at which to purchase the Reserved Instance, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ).

$sel:limitPrice:PurchaseReservedInstancesOffering', purchaseReservedInstancesOffering_limitPrice - Specified for Reserved Instance Marketplace offerings to limit the total order and ensure that the Reserved Instances are not purchased at unexpected prices.

$sel:instanceCount:PurchaseReservedInstancesOffering', purchaseReservedInstancesOffering_instanceCount - The number of Reserved Instances to purchase.

$sel:reservedInstancesOfferingId:PurchaseReservedInstancesOffering', purchaseReservedInstancesOffering_reservedInstancesOfferingId - The ID of the Reserved Instance offering to purchase.

Request Lenses

purchaseReservedInstancesOffering_dryRun :: Lens' PurchaseReservedInstancesOffering (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

purchaseReservedInstancesOffering_purchaseTime :: Lens' PurchaseReservedInstancesOffering (Maybe UTCTime)

The time at which to purchase the Reserved Instance, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ).

purchaseReservedInstancesOffering_limitPrice :: Lens' PurchaseReservedInstancesOffering (Maybe ReservedInstanceLimitPrice)

Specified for Reserved Instance Marketplace offerings to limit the total order and ensure that the Reserved Instances are not purchased at unexpected prices.

purchaseReservedInstancesOffering_instanceCount :: Lens' PurchaseReservedInstancesOffering Int

The number of Reserved Instances to purchase.

purchaseReservedInstancesOffering_reservedInstancesOfferingId :: Lens' PurchaseReservedInstancesOffering Text

The ID of the Reserved Instance offering to purchase.

Destructuring the Response

data PurchaseReservedInstancesOfferingResponse

Contains the output of PurchaseReservedInstancesOffering.

See: newPurchaseReservedInstancesOfferingResponse smart constructor.

Constructors

PurchaseReservedInstancesOfferingResponse'

Fields

reservedInstancesId :: Maybe Text

The IDs of the purchased Reserved Instances. If your purchase crosses into a discounted pricing tier, the final Reserved Instances IDs might change. For more information, see Crossing pricing tiers in the Amazon Elastic Compute Cloud User Guide.

httpStatus :: Int

The response's http status code.

newPurchaseReservedInstancesOfferingResponse

Arguments

:: Int
$sel:httpStatus:PurchaseReservedInstancesOfferingResponse'

-> PurchaseReservedInstancesOfferingResponse

Create a value of PurchaseReservedInstancesOfferingResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reservedInstancesId:PurchaseReservedInstancesOfferingResponse', purchaseReservedInstancesOfferingResponse_reservedInstancesId - The IDs of the purchased Reserved Instances. If your purchase crosses into a discounted pricing tier, the final Reserved Instances IDs might change. For more information, see Crossing pricing tiers in the Amazon Elastic Compute Cloud User Guide.

$sel:httpStatus:PurchaseReservedInstancesOfferingResponse', purchaseReservedInstancesOfferingResponse_httpStatus - The response's http status code.

Response Lenses

purchaseReservedInstancesOfferingResponse_reservedInstancesId :: Lens' PurchaseReservedInstancesOfferingResponse (Maybe Text)

The IDs of the purchased Reserved Instances. If your purchase crosses into a discounted pricing tier, the final Reserved Instances IDs might change. For more information, see Crossing pricing tiers in the Amazon Elastic Compute Cloud User Guide.

purchaseReservedInstancesOfferingResponse_httpStatus :: Lens' PurchaseReservedInstancesOfferingResponse Int

The response's http status code.