Network.AWS.EC2.RegisterImage

Description

Registers an AMI. When you're creating an AMI, this is the final step you must complete before you can launch an instance from the AMI. For more information about creating AMIs, see Creating your own AMIs in the Amazon Elastic Compute Cloud User Guide.

For Amazon EBS-backed instances, CreateImage creates and registers the AMI in a single request, so you don't have to register the AMI yourself.

If needed, you can deregister an AMI at any time. Any modifications you make to an AMI backed by an instance store volume invalidates its registration. If you make changes to an image, deregister the previous image and register the new image.

Register a snapshot of a root device volume

You can use RegisterImage to create an Amazon EBS-backed Linux AMI from a snapshot of a root device volume. You specify the snapshot using a block device mapping. You can't set the encryption state of the volume using the block device mapping. If the snapshot is encrypted, or encryption by default is enabled, the root volume of an instance launched from the AMI is encrypted.

For more information, see Create a Linux AMI from a snapshot and Use encryption with Amazon EBS-backed AMIs in the Amazon Elastic Compute Cloud User Guide.

Amazon Web Services Marketplace product codes

If any snapshots have Amazon Web Services Marketplace product codes, they are copied to the new AMI.

Windows and some Linux distributions, such as Red Hat Enterprise Linux (RHEL) and SUSE Linux Enterprise Server (SLES), use the Amazon EC2 billing product code associated with an AMI to verify the subscription status for package updates. To create a new AMI for operating systems that require a billing product code, instead of registering the AMI, do the following to preserve the billing product code association:

Launch an instance from an existing AMI with that billing product code.

Customize the instance.

Create an AMI from the instance using CreateImage.

If you purchase a Reserved Instance to apply to an On-Demand Instance that was launched from an AMI with a billing product code, make sure that the Reserved Instance has the matching billing product code. If you purchase a Reserved Instance without the matching billing product code, the Reserved Instance will not be applied to the On-Demand Instance. For information about how to obtain the platform details and billing information of an AMI, see Understanding AMI billing in the Amazon Elastic Compute Cloud User Guide.

Creating a Request

data RegisterImage

Contains the parameters for RegisterImage.

See: newRegisterImage smart constructor.

Constructors

RegisterImage'

Fields

virtualizationType :: Maybe Text

The type of virtualization (hvm | paravirtual).

Default: paravirtual

rootDeviceName :: Maybe Text

The device name of the root device volume (for example, /dev/sda1).

ramdiskId :: Maybe Text

The ID of the RAM disk.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

architecture :: Maybe ArchitectureValues

The architecture of the AMI.

Default: For Amazon EBS-backed AMIs, i386. For instance store-backed AMIs, the architecture specified in the manifest file.

bootMode :: Maybe BootModeValues

The boot mode of the AMI. For more information, see Boot modes in the Amazon Elastic Compute Cloud User Guide.

sriovNetSupport :: Maybe Text

Set to simple to enable enhanced networking with the Intel 82599 Virtual Function interface for the AMI and any instances that you launch from the AMI.

There is no way to disable sriovNetSupport at this time.

This option is supported only for HVM AMIs. Specifying this option with a PV AMI can make instances launched from the AMI unreachable.

blockDeviceMappings :: Maybe [BlockDeviceMapping]

The block device mapping entries.

If you specify an Amazon EBS volume using the ID of an Amazon EBS snapshot, you can't specify the encryption state of the volume.

If you create an AMI on an Outpost, then all backing snapshots must be on the same Outpost or in the Region of that Outpost. AMIs on an Outpost that include local snapshots can be used to launch instances on the same Outpost only. For more information, Amazon EBS local snapshots on Outposts in the Amazon Elastic Compute Cloud User Guide.

description :: Maybe Text

A description for your AMI.

kernelId :: Maybe Text

The ID of the kernel.

billingProducts :: Maybe [Text]

The billing product codes. Your account must be authorized to specify billing product codes. Otherwise, you can use the Amazon Web Services Marketplace to bill for the use of an AMI.

enaSupport :: Maybe Bool

Set to true to enable enhanced networking with ENA for the AMI and any instances that you launch from the AMI.

This option is supported only for HVM AMIs. Specifying this option with a PV AMI can make instances launched from the AMI unreachable.

imageLocation :: Maybe Text

The full path to your AMI manifest in Amazon S3 storage. The specified bucket must have the aws-exec-read canned access control list (ACL) to ensure that it can be accessed by Amazon EC2. For more information, see Canned ACLs in the Amazon S3 Service Developer Guide.

name :: Text

A name for your AMI.

Constraints: 3-128 alphanumeric characters, parentheses (()), square brackets ([]), spaces ( ), periods (.), slashes (/), dashes (-), single quotes ('), at-signs (@), or underscores(_)

newRegisterImage

Arguments

:: Text
$sel:name:RegisterImage'

-> RegisterImage

Create a value of RegisterImage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:virtualizationType:RegisterImage', registerImage_virtualizationType - The type of virtualization (hvm | paravirtual).

Default: paravirtual

$sel:rootDeviceName:RegisterImage', registerImage_rootDeviceName - The device name of the root device volume (for example, /dev/sda1).

$sel:ramdiskId:RegisterImage', registerImage_ramdiskId - The ID of the RAM disk.

$sel:dryRun:RegisterImage', registerImage_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:architecture:RegisterImage', registerImage_architecture - The architecture of the AMI.

Default: For Amazon EBS-backed AMIs, i386. For instance store-backed AMIs, the architecture specified in the manifest file.

$sel:bootMode:RegisterImage', registerImage_bootMode - The boot mode of the AMI. For more information, see Boot modes in the Amazon Elastic Compute Cloud User Guide.

$sel:sriovNetSupport:RegisterImage', registerImage_sriovNetSupport - Set to simple to enable enhanced networking with the Intel 82599 Virtual Function interface for the AMI and any instances that you launch from the AMI.

There is no way to disable sriovNetSupport at this time.

This option is supported only for HVM AMIs. Specifying this option with a PV AMI can make instances launched from the AMI unreachable.

$sel:blockDeviceMappings:RegisterImage', registerImage_blockDeviceMappings - The block device mapping entries.

If you specify an Amazon EBS volume using the ID of an Amazon EBS snapshot, you can't specify the encryption state of the volume.

If you create an AMI on an Outpost, then all backing snapshots must be on the same Outpost or in the Region of that Outpost. AMIs on an Outpost that include local snapshots can be used to launch instances on the same Outpost only. For more information, Amazon EBS local snapshots on Outposts in the Amazon Elastic Compute Cloud User Guide.

$sel:description:RegisterImage', registerImage_description - A description for your AMI.

$sel:kernelId:RegisterImage', registerImage_kernelId - The ID of the kernel.

$sel:billingProducts:RegisterImage', registerImage_billingProducts - The billing product codes. Your account must be authorized to specify billing product codes. Otherwise, you can use the Amazon Web Services Marketplace to bill for the use of an AMI.

$sel:enaSupport:RegisterImage', registerImage_enaSupport - Set to true to enable enhanced networking with ENA for the AMI and any instances that you launch from the AMI.

This option is supported only for HVM AMIs. Specifying this option with a PV AMI can make instances launched from the AMI unreachable.

$sel:imageLocation:RegisterImage', registerImage_imageLocation - The full path to your AMI manifest in Amazon S3 storage. The specified bucket must have the aws-exec-read canned access control list (ACL) to ensure that it can be accessed by Amazon EC2. For more information, see Canned ACLs in the Amazon S3 Service Developer Guide.

$sel:name:RegisterImage', registerImage_name - A name for your AMI.

Constraints: 3-128 alphanumeric characters, parentheses (()), square brackets ([]), spaces ( ), periods (.), slashes (/), dashes (-), single quotes ('), at-signs (@), or underscores(_)

Request Lenses

registerImage_virtualizationType :: Lens' RegisterImage (Maybe Text)

The type of virtualization (hvm | paravirtual).

Default: paravirtual

registerImage_rootDeviceName :: Lens' RegisterImage (Maybe Text)

The device name of the root device volume (for example, /dev/sda1).

registerImage_ramdiskId :: Lens' RegisterImage (Maybe Text)

The ID of the RAM disk.

registerImage_dryRun :: Lens' RegisterImage (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

registerImage_architecture :: Lens' RegisterImage (Maybe ArchitectureValues)

The architecture of the AMI.

Default: For Amazon EBS-backed AMIs, i386. For instance store-backed AMIs, the architecture specified in the manifest file.

registerImage_bootMode :: Lens' RegisterImage (Maybe BootModeValues)

The boot mode of the AMI. For more information, see Boot modes in the Amazon Elastic Compute Cloud User Guide.

registerImage_sriovNetSupport :: Lens' RegisterImage (Maybe Text)

Set to simple to enable enhanced networking with the Intel 82599 Virtual Function interface for the AMI and any instances that you launch from the AMI.

There is no way to disable sriovNetSupport at this time.

This option is supported only for HVM AMIs. Specifying this option with a PV AMI can make instances launched from the AMI unreachable.

registerImage_blockDeviceMappings :: Lens' RegisterImage (Maybe [BlockDeviceMapping])

The block device mapping entries.

If you specify an Amazon EBS volume using the ID of an Amazon EBS snapshot, you can't specify the encryption state of the volume.

If you create an AMI on an Outpost, then all backing snapshots must be on the same Outpost or in the Region of that Outpost. AMIs on an Outpost that include local snapshots can be used to launch instances on the same Outpost only. For more information, Amazon EBS local snapshots on Outposts in the Amazon Elastic Compute Cloud User Guide.

registerImage_description :: Lens' RegisterImage (Maybe Text)

A description for your AMI.

registerImage_kernelId :: Lens' RegisterImage (Maybe Text)

The ID of the kernel.

registerImage_billingProducts :: Lens' RegisterImage (Maybe [Text])

The billing product codes. Your account must be authorized to specify billing product codes. Otherwise, you can use the Amazon Web Services Marketplace to bill for the use of an AMI.

registerImage_enaSupport :: Lens' RegisterImage (Maybe Bool)

Set to true to enable enhanced networking with ENA for the AMI and any instances that you launch from the AMI.

This option is supported only for HVM AMIs. Specifying this option with a PV AMI can make instances launched from the AMI unreachable.

registerImage_imageLocation :: Lens' RegisterImage (Maybe Text)

The full path to your AMI manifest in Amazon S3 storage. The specified bucket must have the aws-exec-read canned access control list (ACL) to ensure that it can be accessed by Amazon EC2. For more information, see Canned ACLs in the Amazon S3 Service Developer Guide.

registerImage_name :: Lens' RegisterImage Text

A name for your AMI.

Constraints: 3-128 alphanumeric characters, parentheses (()), square brackets ([]), spaces ( ), periods (.), slashes (/), dashes (-), single quotes ('), at-signs (@), or underscores(_)

Destructuring the Response

data RegisterImageResponse

Contains the output of RegisterImage.

See: newRegisterImageResponse smart constructor.

Constructors

RegisterImageResponse'

Fields

imageId :: Maybe Text

The ID of the newly registered AMI.

httpStatus :: Int

The response's http status code.

newRegisterImageResponse

Arguments

:: Int
$sel:httpStatus:RegisterImageResponse'

-> RegisterImageResponse

Create a value of RegisterImageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:imageId:RegisterImageResponse', registerImageResponse_imageId - The ID of the newly registered AMI.

$sel:httpStatus:RegisterImageResponse', registerImageResponse_httpStatus - The response's http status code.

Response Lenses

registerImageResponse_imageId :: Lens' RegisterImageResponse (Maybe Text)

The ID of the newly registered AMI.

registerImageResponse_httpStatus :: Lens' RegisterImageResponse Int

The response's http status code.