Network.AWS.EC2.RegisterInstanceEventNotificationAttributes

Description

Registers a set of tag keys to include in scheduled event notifications for your resources.

To remove tags, use .

Creating a Request

data RegisterInstanceEventNotificationAttributes

See: newRegisterInstanceEventNotificationAttributes smart constructor.

Constructors

RegisterInstanceEventNotificationAttributes'

Fields

instanceTagAttribute :: Maybe RegisterInstanceTagAttributeRequest

Information about the tag keys to register.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

newRegisterInstanceEventNotificationAttributes :: RegisterInstanceEventNotificationAttributes

Create a value of RegisterInstanceEventNotificationAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceTagAttribute:RegisterInstanceEventNotificationAttributes', registerInstanceEventNotificationAttributes_instanceTagAttribute - Information about the tag keys to register.

$sel:dryRun:RegisterInstanceEventNotificationAttributes', registerInstanceEventNotificationAttributes_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

Request Lenses

registerInstanceEventNotificationAttributes_instanceTagAttribute :: Lens' RegisterInstanceEventNotificationAttributes (Maybe RegisterInstanceTagAttributeRequest)

Information about the tag keys to register.

registerInstanceEventNotificationAttributes_dryRun :: Lens' RegisterInstanceEventNotificationAttributes (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

Destructuring the Response

data RegisterInstanceEventNotificationAttributesResponse

See: newRegisterInstanceEventNotificationAttributesResponse smart constructor.

Constructors

RegisterInstanceEventNotificationAttributesResponse'

Fields

instanceTagAttribute :: Maybe InstanceTagNotificationAttribute

The resulting set of tag keys.

httpStatus :: Int

The response's http status code.

newRegisterInstanceEventNotificationAttributesResponse

Arguments

:: Int
$sel:httpStatus:RegisterInstanceEventNotificationAttributesResponse'

-> RegisterInstanceEventNotificationAttributesResponse

Create a value of RegisterInstanceEventNotificationAttributesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceTagAttribute:RegisterInstanceEventNotificationAttributes', registerInstanceEventNotificationAttributesResponse_instanceTagAttribute - The resulting set of tag keys.

$sel:httpStatus:RegisterInstanceEventNotificationAttributesResponse', registerInstanceEventNotificationAttributesResponse_httpStatus - The response's http status code.

Response Lenses

registerInstanceEventNotificationAttributesResponse_instanceTagAttribute :: Lens' RegisterInstanceEventNotificationAttributesResponse (Maybe InstanceTagNotificationAttribute)

The resulting set of tag keys.

registerInstanceEventNotificationAttributesResponse_httpStatus :: Lens' RegisterInstanceEventNotificationAttributesResponse Int

The response's http status code.