Network.AWS.EC2.RegisterTransitGatewayMulticastGroupMembers

Description

Registers members (network interfaces) with the transit gateway multicast group. A member is a network interface associated with a supported EC2 instance that receives multicast traffic. For information about supported instances, see Multicast Consideration in Amazon VPC Transit Gateways.

After you add the members, use SearchTransitGatewayMulticastGroups to verify that the members were added to the transit gateway multicast group.

Creating a Request

data RegisterTransitGatewayMulticastGroupMembers

See: newRegisterTransitGatewayMulticastGroupMembers smart constructor.

Constructors

RegisterTransitGatewayMulticastGroupMembers'

Fields

transitGatewayMulticastDomainId :: Maybe Text

The ID of the transit gateway multicast domain.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

networkInterfaceIds :: Maybe [Text]

The group members' network interface IDs to register with the transit gateway multicast group.

groupIpAddress :: Maybe Text

The IP address assigned to the transit gateway multicast group.

newRegisterTransitGatewayMulticastGroupMembers :: RegisterTransitGatewayMulticastGroupMembers

Create a value of RegisterTransitGatewayMulticastGroupMembers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:transitGatewayMulticastDomainId:RegisterTransitGatewayMulticastGroupMembers', registerTransitGatewayMulticastGroupMembers_transitGatewayMulticastDomainId - The ID of the transit gateway multicast domain.

$sel:dryRun:RegisterTransitGatewayMulticastGroupMembers', registerTransitGatewayMulticastGroupMembers_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:networkInterfaceIds:RegisterTransitGatewayMulticastGroupMembers', registerTransitGatewayMulticastGroupMembers_networkInterfaceIds - The group members' network interface IDs to register with the transit gateway multicast group.

$sel:groupIpAddress:RegisterTransitGatewayMulticastGroupMembers', registerTransitGatewayMulticastGroupMembers_groupIpAddress - The IP address assigned to the transit gateway multicast group.

Request Lenses

registerTransitGatewayMulticastGroupMembers_transitGatewayMulticastDomainId :: Lens' RegisterTransitGatewayMulticastGroupMembers (Maybe Text)

The ID of the transit gateway multicast domain.

registerTransitGatewayMulticastGroupMembers_dryRun :: Lens' RegisterTransitGatewayMulticastGroupMembers (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

registerTransitGatewayMulticastGroupMembers_networkInterfaceIds :: Lens' RegisterTransitGatewayMulticastGroupMembers (Maybe [Text])

The group members' network interface IDs to register with the transit gateway multicast group.

registerTransitGatewayMulticastGroupMembers_groupIpAddress :: Lens' RegisterTransitGatewayMulticastGroupMembers (Maybe Text)

The IP address assigned to the transit gateway multicast group.

Destructuring the Response

data RegisterTransitGatewayMulticastGroupMembersResponse

See: newRegisterTransitGatewayMulticastGroupMembersResponse smart constructor.

Constructors

RegisterTransitGatewayMulticastGroupMembersResponse'

Fields

registeredMulticastGroupMembers :: Maybe TransitGatewayMulticastRegisteredGroupMembers

Information about the registered transit gateway multicast group members.

httpStatus :: Int

The response's http status code.

newRegisterTransitGatewayMulticastGroupMembersResponse

Arguments

:: Int
$sel:httpStatus:RegisterTransitGatewayMulticastGroupMembersResponse'

-> RegisterTransitGatewayMulticastGroupMembersResponse

Create a value of RegisterTransitGatewayMulticastGroupMembersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:registeredMulticastGroupMembers:RegisterTransitGatewayMulticastGroupMembersResponse', registerTransitGatewayMulticastGroupMembersResponse_registeredMulticastGroupMembers - Information about the registered transit gateway multicast group members.

$sel:httpStatus:RegisterTransitGatewayMulticastGroupMembersResponse', registerTransitGatewayMulticastGroupMembersResponse_httpStatus - The response's http status code.

Response Lenses

registerTransitGatewayMulticastGroupMembersResponse_registeredMulticastGroupMembers :: Lens' RegisterTransitGatewayMulticastGroupMembersResponse (Maybe TransitGatewayMulticastRegisteredGroupMembers)

Information about the registered transit gateway multicast group members.

registerTransitGatewayMulticastGroupMembersResponse_httpStatus :: Lens' RegisterTransitGatewayMulticastGroupMembersResponse Int

The response's http status code.