Network.AWS.EC2.RejectTransitGatewayMulticastDomainAssociations

Description

Rejects a request to associate cross-account subnets with a transit gateway multicast domain.

Creating a Request

data RejectTransitGatewayMulticastDomainAssociations

See: newRejectTransitGatewayMulticastDomainAssociations smart constructor.

Constructors

RejectTransitGatewayMulticastDomainAssociations'

Fields

transitGatewayMulticastDomainId :: Maybe Text

The ID of the transit gateway multicast domain.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

subnetIds :: Maybe [Text]

The IDs of the subnets to associate with the transit gateway multicast domain.

transitGatewayAttachmentId :: Maybe Text

The ID of the transit gateway attachment.

newRejectTransitGatewayMulticastDomainAssociations :: RejectTransitGatewayMulticastDomainAssociations

Create a value of RejectTransitGatewayMulticastDomainAssociations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:transitGatewayMulticastDomainId:RejectTransitGatewayMulticastDomainAssociations', rejectTransitGatewayMulticastDomainAssociations_transitGatewayMulticastDomainId - The ID of the transit gateway multicast domain.

$sel:dryRun:RejectTransitGatewayMulticastDomainAssociations', rejectTransitGatewayMulticastDomainAssociations_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:subnetIds:RejectTransitGatewayMulticastDomainAssociations', rejectTransitGatewayMulticastDomainAssociations_subnetIds - The IDs of the subnets to associate with the transit gateway multicast domain.

$sel:transitGatewayAttachmentId:RejectTransitGatewayMulticastDomainAssociations', rejectTransitGatewayMulticastDomainAssociations_transitGatewayAttachmentId - The ID of the transit gateway attachment.

Request Lenses

rejectTransitGatewayMulticastDomainAssociations_transitGatewayMulticastDomainId :: Lens' RejectTransitGatewayMulticastDomainAssociations (Maybe Text)

The ID of the transit gateway multicast domain.

rejectTransitGatewayMulticastDomainAssociations_dryRun :: Lens' RejectTransitGatewayMulticastDomainAssociations (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

rejectTransitGatewayMulticastDomainAssociations_subnetIds :: Lens' RejectTransitGatewayMulticastDomainAssociations (Maybe [Text])

The IDs of the subnets to associate with the transit gateway multicast domain.

rejectTransitGatewayMulticastDomainAssociations_transitGatewayAttachmentId :: Lens' RejectTransitGatewayMulticastDomainAssociations (Maybe Text)

The ID of the transit gateway attachment.

Destructuring the Response

data RejectTransitGatewayMulticastDomainAssociationsResponse

See: newRejectTransitGatewayMulticastDomainAssociationsResponse smart constructor.

Constructors

RejectTransitGatewayMulticastDomainAssociationsResponse'

Fields

associations :: Maybe TransitGatewayMulticastDomainAssociations

httpStatus :: Int

The response's http status code.

newRejectTransitGatewayMulticastDomainAssociationsResponse

Arguments

:: Int
$sel:httpStatus:RejectTransitGatewayMulticastDomainAssociationsResponse'

-> RejectTransitGatewayMulticastDomainAssociationsResponse

Create a value of RejectTransitGatewayMulticastDomainAssociationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:associations:RejectTransitGatewayMulticastDomainAssociationsResponse', rejectTransitGatewayMulticastDomainAssociationsResponse_associations - Undocumented member.

$sel:httpStatus:RejectTransitGatewayMulticastDomainAssociationsResponse', rejectTransitGatewayMulticastDomainAssociationsResponse_httpStatus - The response's http status code.

Response Lenses

rejectTransitGatewayMulticastDomainAssociationsResponse_associations :: Lens' RejectTransitGatewayMulticastDomainAssociationsResponse (Maybe TransitGatewayMulticastDomainAssociations)

Undocumented member.

rejectTransitGatewayMulticastDomainAssociationsResponse_httpStatus :: Lens' RejectTransitGatewayMulticastDomainAssociationsResponse Int

The response's http status code.