Network.AWS.EC2.ReplaceNetworkAclEntry

Description

Replaces an entry (rule) in a network ACL. For more information, see Network ACLs in the Amazon Virtual Private Cloud User Guide.

Creating a Request

data ReplaceNetworkAclEntry

See: newReplaceNetworkAclEntry smart constructor.

Constructors

ReplaceNetworkAclEntry'

Fields

portRange :: Maybe PortRange

TCP or UDP protocols: The range of ports the rule applies to. Required if specifying protocol 6 (TCP) or 17 (UDP).

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

icmpTypeCode :: Maybe IcmpTypeCode

ICMP protocol: The ICMP or ICMPv6 type and code. Required if specifying protocol 1 (ICMP) or protocol 58 (ICMPv6) with an IPv6 CIDR block.

ipv6CidrBlock :: Maybe Text

The IPv6 network range to allow or deny, in CIDR notation (for example 2001:bd8:1234:1a00::/64).

cidrBlock :: Maybe Text

The IPv4 network range to allow or deny, in CIDR notation (for example 172.16.0.0/24).

egress :: Bool

Indicates whether to replace the egress rule.

Default: If no value is specified, we replace the ingress rule.

networkAclId :: Text

The ID of the ACL.

protocol :: Text

The protocol number. A value of "-1" means all protocols. If you specify "-1" or a protocol number other than "6" (TCP), "17" (UDP), or "1" (ICMP), traffic on all ports is allowed, regardless of any ports or ICMP types or codes that you specify. If you specify protocol "58" (ICMPv6) and specify an IPv4 CIDR block, traffic for all ICMP types and codes allowed, regardless of any that you specify. If you specify protocol "58" (ICMPv6) and specify an IPv6 CIDR block, you must specify an ICMP type and code.

ruleAction :: RuleAction

Indicates whether to allow or deny the traffic that matches the rule.

ruleNumber :: Int

The rule number of the entry to replace.

newReplaceNetworkAclEntry

Arguments

:: Bool
$sel:egress:ReplaceNetworkAclEntry'

-> Text
$sel:networkAclId:ReplaceNetworkAclEntry'

-> Text
$sel:protocol:ReplaceNetworkAclEntry'

-> RuleAction
$sel:ruleAction:ReplaceNetworkAclEntry'

-> Int
$sel:ruleNumber:ReplaceNetworkAclEntry'

-> ReplaceNetworkAclEntry

Create a value of ReplaceNetworkAclEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:portRange:ReplaceNetworkAclEntry', replaceNetworkAclEntry_portRange - TCP or UDP protocols: The range of ports the rule applies to. Required if specifying protocol 6 (TCP) or 17 (UDP).

$sel:dryRun:ReplaceNetworkAclEntry', replaceNetworkAclEntry_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:icmpTypeCode:ReplaceNetworkAclEntry', replaceNetworkAclEntry_icmpTypeCode - ICMP protocol: The ICMP or ICMPv6 type and code. Required if specifying protocol 1 (ICMP) or protocol 58 (ICMPv6) with an IPv6 CIDR block.

$sel:ipv6CidrBlock:ReplaceNetworkAclEntry', replaceNetworkAclEntry_ipv6CidrBlock - The IPv6 network range to allow or deny, in CIDR notation (for example 2001:bd8:1234:1a00::/64).

$sel:cidrBlock:ReplaceNetworkAclEntry', replaceNetworkAclEntry_cidrBlock - The IPv4 network range to allow or deny, in CIDR notation (for example 172.16.0.0/24).

$sel:egress:ReplaceNetworkAclEntry', replaceNetworkAclEntry_egress - Indicates whether to replace the egress rule.

Default: If no value is specified, we replace the ingress rule.

$sel:networkAclId:ReplaceNetworkAclEntry', replaceNetworkAclEntry_networkAclId - The ID of the ACL.

$sel:protocol:ReplaceNetworkAclEntry', replaceNetworkAclEntry_protocol - The protocol number. A value of "-1" means all protocols. If you specify "-1" or a protocol number other than "6" (TCP), "17" (UDP), or "1" (ICMP), traffic on all ports is allowed, regardless of any ports or ICMP types or codes that you specify. If you specify protocol "58" (ICMPv6) and specify an IPv4 CIDR block, traffic for all ICMP types and codes allowed, regardless of any that you specify. If you specify protocol "58" (ICMPv6) and specify an IPv6 CIDR block, you must specify an ICMP type and code.

$sel:ruleAction:ReplaceNetworkAclEntry', replaceNetworkAclEntry_ruleAction - Indicates whether to allow or deny the traffic that matches the rule.

$sel:ruleNumber:ReplaceNetworkAclEntry', replaceNetworkAclEntry_ruleNumber - The rule number of the entry to replace.

Request Lenses

replaceNetworkAclEntry_portRange :: Lens' ReplaceNetworkAclEntry (Maybe PortRange)

TCP or UDP protocols: The range of ports the rule applies to. Required if specifying protocol 6 (TCP) or 17 (UDP).

replaceNetworkAclEntry_dryRun :: Lens' ReplaceNetworkAclEntry (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

replaceNetworkAclEntry_icmpTypeCode :: Lens' ReplaceNetworkAclEntry (Maybe IcmpTypeCode)

ICMP protocol: The ICMP or ICMPv6 type and code. Required if specifying protocol 1 (ICMP) or protocol 58 (ICMPv6) with an IPv6 CIDR block.

replaceNetworkAclEntry_ipv6CidrBlock :: Lens' ReplaceNetworkAclEntry (Maybe Text)

The IPv6 network range to allow or deny, in CIDR notation (for example 2001:bd8:1234:1a00::/64).

replaceNetworkAclEntry_cidrBlock :: Lens' ReplaceNetworkAclEntry (Maybe Text)

The IPv4 network range to allow or deny, in CIDR notation (for example 172.16.0.0/24).

replaceNetworkAclEntry_egress :: Lens' ReplaceNetworkAclEntry Bool

Indicates whether to replace the egress rule.

Default: If no value is specified, we replace the ingress rule.

replaceNetworkAclEntry_networkAclId :: Lens' ReplaceNetworkAclEntry Text

The ID of the ACL.

replaceNetworkAclEntry_protocol :: Lens' ReplaceNetworkAclEntry Text

The protocol number. A value of "-1" means all protocols. If you specify "-1" or a protocol number other than "6" (TCP), "17" (UDP), or "1" (ICMP), traffic on all ports is allowed, regardless of any ports or ICMP types or codes that you specify. If you specify protocol "58" (ICMPv6) and specify an IPv4 CIDR block, traffic for all ICMP types and codes allowed, regardless of any that you specify. If you specify protocol "58" (ICMPv6) and specify an IPv6 CIDR block, you must specify an ICMP type and code.

replaceNetworkAclEntry_ruleAction :: Lens' ReplaceNetworkAclEntry RuleAction

Indicates whether to allow or deny the traffic that matches the rule.

replaceNetworkAclEntry_ruleNumber :: Lens' ReplaceNetworkAclEntry Int

The rule number of the entry to replace.

Destructuring the Response

data ReplaceNetworkAclEntryResponse

See: newReplaceNetworkAclEntryResponse smart constructor.

Constructors

ReplaceNetworkAclEntryResponse'

newReplaceNetworkAclEntryResponse :: ReplaceNetworkAclEntryResponse

Create a value of ReplaceNetworkAclEntryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.