Network.AWS.EC2.ReplaceRoute

Description

Replaces an existing route within a route table in a VPC. You must provide only one of the following: internet gateway, virtual private gateway, NAT instance, NAT gateway, VPC peering connection, network interface, egress-only internet gateway, or transit gateway.

For more information, see Route tables in the Amazon Virtual Private Cloud User Guide.

Creating a Request

data ReplaceRoute

See: newReplaceRoute smart constructor.

Constructors

ReplaceRoute'

Fields

instanceId :: Maybe Text

The ID of a NAT instance in your VPC.

vpcPeeringConnectionId :: Maybe Text

The ID of a VPC peering connection.

localTarget :: Maybe Bool

Specifies whether to reset the local route to its default target (local).

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

vpcEndpointId :: Maybe Text

The ID of a VPC endpoint. Supported for Gateway Load Balancer endpoints only.

destinationPrefixListId :: Maybe Text

The ID of the prefix list for the route.

destinationIpv6CidrBlock :: Maybe Text

The IPv6 CIDR address block used for the destination match. The value that you provide must match the CIDR of an existing route in the table.

localGatewayId :: Maybe Text

The ID of the local gateway.

egressOnlyInternetGatewayId :: Maybe Text

IPv6 traffic only

The ID of an egress-only internet gateway.

destinationCidrBlock :: Maybe Text

The IPv4 CIDR address block used for the destination match. The value that you provide must match the CIDR of an existing route in the table.

carrierGatewayId :: Maybe Text

IPv4 traffic only

The ID of a carrier gateway.

natGatewayId :: Maybe Text

IPv4 traffic only

The ID of a NAT gateway.

networkInterfaceId :: Maybe Text

The ID of a network interface.

gatewayId :: Maybe Text

The ID of an internet gateway or virtual private gateway.

transitGatewayId :: Maybe Text

The ID of a transit gateway.

routeTableId :: Text

The ID of the route table.

newReplaceRoute

Arguments

:: Text
$sel:routeTableId:ReplaceRoute'

-> ReplaceRoute

Create a value of ReplaceRoute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceId:ReplaceRoute', replaceRoute_instanceId - The ID of a NAT instance in your VPC.

$sel:vpcPeeringConnectionId:ReplaceRoute', replaceRoute_vpcPeeringConnectionId - The ID of a VPC peering connection.

$sel:localTarget:ReplaceRoute', replaceRoute_localTarget - Specifies whether to reset the local route to its default target (local).

$sel:dryRun:ReplaceRoute', replaceRoute_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:vpcEndpointId:ReplaceRoute', replaceRoute_vpcEndpointId - The ID of a VPC endpoint. Supported for Gateway Load Balancer endpoints only.

$sel:destinationPrefixListId:ReplaceRoute', replaceRoute_destinationPrefixListId - The ID of the prefix list for the route.

$sel:destinationIpv6CidrBlock:ReplaceRoute', replaceRoute_destinationIpv6CidrBlock - The IPv6 CIDR address block used for the destination match. The value that you provide must match the CIDR of an existing route in the table.

$sel:localGatewayId:ReplaceRoute', replaceRoute_localGatewayId - The ID of the local gateway.

$sel:egressOnlyInternetGatewayId:ReplaceRoute', replaceRoute_egressOnlyInternetGatewayId - [IPv6 traffic only] The ID of an egress-only internet gateway.

$sel:destinationCidrBlock:ReplaceRoute', replaceRoute_destinationCidrBlock - The IPv4 CIDR address block used for the destination match. The value that you provide must match the CIDR of an existing route in the table.

$sel:carrierGatewayId:ReplaceRoute', replaceRoute_carrierGatewayId - [IPv4 traffic only] The ID of a carrier gateway.

$sel:natGatewayId:ReplaceRoute', replaceRoute_natGatewayId - [IPv4 traffic only] The ID of a NAT gateway.

$sel:networkInterfaceId:ReplaceRoute', replaceRoute_networkInterfaceId - The ID of a network interface.

$sel:gatewayId:ReplaceRoute', replaceRoute_gatewayId - The ID of an internet gateway or virtual private gateway.

$sel:transitGatewayId:ReplaceRoute', replaceRoute_transitGatewayId - The ID of a transit gateway.

$sel:routeTableId:ReplaceRoute', replaceRoute_routeTableId - The ID of the route table.

Request Lenses

replaceRoute_instanceId :: Lens' ReplaceRoute (Maybe Text)

The ID of a NAT instance in your VPC.

replaceRoute_vpcPeeringConnectionId :: Lens' ReplaceRoute (Maybe Text)

The ID of a VPC peering connection.

replaceRoute_localTarget :: Lens' ReplaceRoute (Maybe Bool)

Specifies whether to reset the local route to its default target (local).

replaceRoute_dryRun :: Lens' ReplaceRoute (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

replaceRoute_vpcEndpointId :: Lens' ReplaceRoute (Maybe Text)

The ID of a VPC endpoint. Supported for Gateway Load Balancer endpoints only.

replaceRoute_destinationPrefixListId :: Lens' ReplaceRoute (Maybe Text)

The ID of the prefix list for the route.

replaceRoute_destinationIpv6CidrBlock :: Lens' ReplaceRoute (Maybe Text)

The IPv6 CIDR address block used for the destination match. The value that you provide must match the CIDR of an existing route in the table.

replaceRoute_localGatewayId :: Lens' ReplaceRoute (Maybe Text)

The ID of the local gateway.

replaceRoute_egressOnlyInternetGatewayId :: Lens' ReplaceRoute (Maybe Text)

IPv6 traffic only

The ID of an egress-only internet gateway.

replaceRoute_destinationCidrBlock :: Lens' ReplaceRoute (Maybe Text)

The IPv4 CIDR address block used for the destination match. The value that you provide must match the CIDR of an existing route in the table.

replaceRoute_carrierGatewayId :: Lens' ReplaceRoute (Maybe Text)

IPv4 traffic only

The ID of a carrier gateway.

replaceRoute_natGatewayId :: Lens' ReplaceRoute (Maybe Text)

IPv4 traffic only

The ID of a NAT gateway.

replaceRoute_networkInterfaceId :: Lens' ReplaceRoute (Maybe Text)

The ID of a network interface.

replaceRoute_gatewayId :: Lens' ReplaceRoute (Maybe Text)

The ID of an internet gateway or virtual private gateway.

replaceRoute_transitGatewayId :: Lens' ReplaceRoute (Maybe Text)

The ID of a transit gateway.

replaceRoute_routeTableId :: Lens' ReplaceRoute Text

The ID of the route table.

Destructuring the Response

data ReplaceRouteResponse

See: newReplaceRouteResponse smart constructor.

Constructors

ReplaceRouteResponse'

newReplaceRouteResponse :: ReplaceRouteResponse

Create a value of ReplaceRouteResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.