Network.AWS.EC2.RequestSpotFleet

Description

Creates a Spot Fleet request.

The Spot Fleet request specifies the total target capacity and the On-Demand target capacity. Amazon EC2 calculates the difference between the total capacity and On-Demand capacity, and launches the difference as Spot capacity.

You can submit a single request that includes multiple launch specifications that vary by instance type, AMI, Availability Zone, or subnet.

By default, the Spot Fleet requests Spot Instances in the Spot Instance pool where the price per unit is the lowest. Each launch specification can include its own instance weighting that reflects the value of the instance type to your application workload.

Alternatively, you can specify that the Spot Fleet distribute the target capacity across the Spot pools included in its launch specifications. By ensuring that the Spot Instances in your Spot Fleet are in different Spot pools, you can improve the availability of your fleet.

You can specify tags for the Spot Fleet request and instances launched by the fleet. You cannot tag other resource types in a Spot Fleet request because only the spot-fleet-request and instance resource types are supported.

For more information, see Spot Fleet requests in the Amazon EC2 User Guide for Linux Instances.

Creating a Request

data RequestSpotFleet

Contains the parameters for RequestSpotFleet.

See: newRequestSpotFleet smart constructor.

Constructors

RequestSpotFleet'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

spotFleetRequestConfig :: SpotFleetRequestConfigData

The configuration for the Spot Fleet request.

newRequestSpotFleet

Arguments

:: SpotFleetRequestConfigData
$sel:spotFleetRequestConfig:RequestSpotFleet'

-> RequestSpotFleet

Create a value of RequestSpotFleet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:RequestSpotFleet', requestSpotFleet_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:spotFleetRequestConfig:RequestSpotFleet', requestSpotFleet_spotFleetRequestConfig - The configuration for the Spot Fleet request.

Request Lenses

requestSpotFleet_dryRun :: Lens' RequestSpotFleet (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

requestSpotFleet_spotFleetRequestConfig :: Lens' RequestSpotFleet SpotFleetRequestConfigData

The configuration for the Spot Fleet request.

Destructuring the Response

data RequestSpotFleetResponse

Contains the output of RequestSpotFleet.

See: newRequestSpotFleetResponse smart constructor.

Constructors

RequestSpotFleetResponse'

Fields

spotFleetRequestId :: Maybe Text

The ID of the Spot Fleet request.

httpStatus :: Int

The response's http status code.

newRequestSpotFleetResponse

Arguments

:: Int
$sel:httpStatus:RequestSpotFleetResponse'

-> RequestSpotFleetResponse

Create a value of RequestSpotFleetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:spotFleetRequestId:RequestSpotFleetResponse', requestSpotFleetResponse_spotFleetRequestId - The ID of the Spot Fleet request.

$sel:httpStatus:RequestSpotFleetResponse', requestSpotFleetResponse_httpStatus - The response's http status code.

Response Lenses

requestSpotFleetResponse_spotFleetRequestId :: Lens' RequestSpotFleetResponse (Maybe Text)

The ID of the Spot Fleet request.

requestSpotFleetResponse_httpStatus :: Lens' RequestSpotFleetResponse Int

The response's http status code.