Network.AWS.EC2.RequestSpotInstances

Description

Creates a Spot Instance request.

For more information, see Spot Instance requests in the Amazon EC2 User Guide for Linux Instances.

Creating a Request

data RequestSpotInstances

Contains the parameters for RequestSpotInstances.

See: newRequestSpotInstances smart constructor.

Constructors

RequestSpotInstances'

Fields

tagSpecifications :: Maybe [TagSpecification]

The key-value pair for tagging the Spot Instance request on creation. The value for ResourceType must be spot-instances-request, otherwise the Spot Instance request fails. To tag the Spot Instance request after it has been created, see CreateTags.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

validFrom :: Maybe ISO8601

The start date of the request. If this is a one-time request, the request becomes active at this date and time and remains active until all instances launch, the request expires, or the request is canceled. If the request is persistent, the request becomes active at this date and time and remains active until it expires or is canceled.

The specified start date and time cannot be equal to the current date and time. You must specify a start date and time that occurs after the current date and time.

spotPrice :: Maybe Text

The maximum price per hour that you are willing to pay for a Spot Instance. The default is the On-Demand price.

launchGroup :: Maybe Text

The instance launch group. Launch groups are Spot Instances that launch together and terminate together.

Default: Instances are launched and terminated individually

blockDurationMinutes :: Maybe Int

Deprecated.

instanceInterruptionBehavior :: Maybe InstanceInterruptionBehavior

The behavior when a Spot Instance is interrupted. The default is terminate.

validUntil :: Maybe ISO8601

The end date of the request, in UTC format (YYYY-MM-DDTHH:MM:SSZ).

For a persistent request, the request remains active until the ValidUntil date and time is reached. Otherwise, the request remains active until you cancel it.

For a one-time request, the request remains active until all instances launch, the request is canceled, or the ValidUntil date and time is reached. By default, the request is valid for 7 days from the date the request was created.

launchSpecification :: Maybe RequestSpotLaunchSpecification

The launch specification.

availabilityZoneGroup :: Maybe Text

The user-specified name for a logical grouping of requests.

When you specify an Availability Zone group in a Spot Instance request, all Spot Instances in the request are launched in the same Availability Zone. Instance proximity is maintained with this parameter, but the choice of Availability Zone is not. The group applies only to requests for Spot Instances of the same instance type. Any additional Spot Instance requests that are specified with the same Availability Zone group name are launched in that same Availability Zone, as long as at least one instance from the group is still active.

If there is no active instance running in the Availability Zone group that you specify for a new Spot Instance request (all instances are terminated, the request is expired, or the maximum price you specified falls below current Spot price), then Amazon EC2 launches the instance in any Availability Zone where the constraint can be met. Consequently, the subsequent set of Spot Instances could be placed in a different zone from the original request, even if you specified the same Availability Zone group.

Default: Instances are launched in any available Availability Zone.

type' :: Maybe SpotInstanceType

The Spot Instance request type.

Default: one-time

instanceCount :: Maybe Int

The maximum number of Spot Instances to launch.

Default: 1

clientToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to Ensure Idempotency in the Amazon EC2 User Guide for Linux Instances.

newRequestSpotInstances :: RequestSpotInstances

Create a value of RequestSpotInstances with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tagSpecifications:RequestSpotInstances', requestSpotInstances_tagSpecifications - The key-value pair for tagging the Spot Instance request on creation. The value for ResourceType must be spot-instances-request, otherwise the Spot Instance request fails. To tag the Spot Instance request after it has been created, see CreateTags.

$sel:dryRun:RequestSpotInstances', requestSpotInstances_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:validFrom:RequestSpotInstances', requestSpotInstances_validFrom - The start date of the request. If this is a one-time request, the request becomes active at this date and time and remains active until all instances launch, the request expires, or the request is canceled. If the request is persistent, the request becomes active at this date and time and remains active until it expires or is canceled.

The specified start date and time cannot be equal to the current date and time. You must specify a start date and time that occurs after the current date and time.

$sel:spotPrice:RequestSpotInstances', requestSpotInstances_spotPrice - The maximum price per hour that you are willing to pay for a Spot Instance. The default is the On-Demand price.

$sel:launchGroup:RequestSpotInstances', requestSpotInstances_launchGroup - The instance launch group. Launch groups are Spot Instances that launch together and terminate together.

Default: Instances are launched and terminated individually

$sel:blockDurationMinutes:RequestSpotInstances', requestSpotInstances_blockDurationMinutes - Deprecated.

$sel:instanceInterruptionBehavior:RequestSpotInstances', requestSpotInstances_instanceInterruptionBehavior - The behavior when a Spot Instance is interrupted. The default is terminate.

$sel:validUntil:RequestSpotInstances', requestSpotInstances_validUntil - The end date of the request, in UTC format (YYYY-MM-DDTHH:MM:SSZ).

For a persistent request, the request remains active until the ValidUntil date and time is reached. Otherwise, the request remains active until you cancel it.

For a one-time request, the request remains active until all instances launch, the request is canceled, or the ValidUntil date and time is reached. By default, the request is valid for 7 days from the date the request was created.

$sel:launchSpecification:RequestSpotInstances', requestSpotInstances_launchSpecification - The launch specification.

$sel:availabilityZoneGroup:RequestSpotInstances', requestSpotInstances_availabilityZoneGroup - The user-specified name for a logical grouping of requests.

When you specify an Availability Zone group in a Spot Instance request, all Spot Instances in the request are launched in the same Availability Zone. Instance proximity is maintained with this parameter, but the choice of Availability Zone is not. The group applies only to requests for Spot Instances of the same instance type. Any additional Spot Instance requests that are specified with the same Availability Zone group name are launched in that same Availability Zone, as long as at least one instance from the group is still active.

If there is no active instance running in the Availability Zone group that you specify for a new Spot Instance request (all instances are terminated, the request is expired, or the maximum price you specified falls below current Spot price), then Amazon EC2 launches the instance in any Availability Zone where the constraint can be met. Consequently, the subsequent set of Spot Instances could be placed in a different zone from the original request, even if you specified the same Availability Zone group.

Default: Instances are launched in any available Availability Zone.

$sel:type':RequestSpotInstances', requestSpotInstances_type - The Spot Instance request type.

Default: one-time

$sel:instanceCount:RequestSpotInstances', requestSpotInstances_instanceCount - The maximum number of Spot Instances to launch.

Default: 1

$sel:clientToken:RequestSpotInstances', requestSpotInstances_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to Ensure Idempotency in the Amazon EC2 User Guide for Linux Instances.

Request Lenses

requestSpotInstances_tagSpecifications :: Lens' RequestSpotInstances (Maybe [TagSpecification])

The key-value pair for tagging the Spot Instance request on creation. The value for ResourceType must be spot-instances-request, otherwise the Spot Instance request fails. To tag the Spot Instance request after it has been created, see CreateTags.

requestSpotInstances_dryRun :: Lens' RequestSpotInstances (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

requestSpotInstances_validFrom :: Lens' RequestSpotInstances (Maybe UTCTime)

The start date of the request. If this is a one-time request, the request becomes active at this date and time and remains active until all instances launch, the request expires, or the request is canceled. If the request is persistent, the request becomes active at this date and time and remains active until it expires or is canceled.

The specified start date and time cannot be equal to the current date and time. You must specify a start date and time that occurs after the current date and time.

requestSpotInstances_spotPrice :: Lens' RequestSpotInstances (Maybe Text)

The maximum price per hour that you are willing to pay for a Spot Instance. The default is the On-Demand price.

requestSpotInstances_launchGroup :: Lens' RequestSpotInstances (Maybe Text)

The instance launch group. Launch groups are Spot Instances that launch together and terminate together.

Default: Instances are launched and terminated individually

requestSpotInstances_blockDurationMinutes :: Lens' RequestSpotInstances (Maybe Int)

Deprecated.

requestSpotInstances_instanceInterruptionBehavior :: Lens' RequestSpotInstances (Maybe InstanceInterruptionBehavior)

The behavior when a Spot Instance is interrupted. The default is terminate.

requestSpotInstances_validUntil :: Lens' RequestSpotInstances (Maybe UTCTime)

The end date of the request, in UTC format (YYYY-MM-DDTHH:MM:SSZ).

For a persistent request, the request remains active until the ValidUntil date and time is reached. Otherwise, the request remains active until you cancel it.

For a one-time request, the request remains active until all instances launch, the request is canceled, or the ValidUntil date and time is reached. By default, the request is valid for 7 days from the date the request was created.

requestSpotInstances_launchSpecification :: Lens' RequestSpotInstances (Maybe RequestSpotLaunchSpecification)

The launch specification.

requestSpotInstances_availabilityZoneGroup :: Lens' RequestSpotInstances (Maybe Text)

The user-specified name for a logical grouping of requests.

When you specify an Availability Zone group in a Spot Instance request, all Spot Instances in the request are launched in the same Availability Zone. Instance proximity is maintained with this parameter, but the choice of Availability Zone is not. The group applies only to requests for Spot Instances of the same instance type. Any additional Spot Instance requests that are specified with the same Availability Zone group name are launched in that same Availability Zone, as long as at least one instance from the group is still active.

If there is no active instance running in the Availability Zone group that you specify for a new Spot Instance request (all instances are terminated, the request is expired, or the maximum price you specified falls below current Spot price), then Amazon EC2 launches the instance in any Availability Zone where the constraint can be met. Consequently, the subsequent set of Spot Instances could be placed in a different zone from the original request, even if you specified the same Availability Zone group.

Default: Instances are launched in any available Availability Zone.

requestSpotInstances_type :: Lens' RequestSpotInstances (Maybe SpotInstanceType)

The Spot Instance request type.

Default: one-time

requestSpotInstances_instanceCount :: Lens' RequestSpotInstances (Maybe Int)

The maximum number of Spot Instances to launch.

Default: 1

requestSpotInstances_clientToken :: Lens' RequestSpotInstances (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to Ensure Idempotency in the Amazon EC2 User Guide for Linux Instances.

Destructuring the Response

data RequestSpotInstancesResponse

Contains the output of RequestSpotInstances.

See: newRequestSpotInstancesResponse smart constructor.

Constructors

RequestSpotInstancesResponse'

Fields

spotInstanceRequests :: Maybe [SpotInstanceRequest]

One or more Spot Instance requests.

httpStatus :: Int

The response's http status code.

newRequestSpotInstancesResponse

Arguments

:: Int
$sel:httpStatus:RequestSpotInstancesResponse'

-> RequestSpotInstancesResponse

Create a value of RequestSpotInstancesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:spotInstanceRequests:RequestSpotInstancesResponse', requestSpotInstancesResponse_spotInstanceRequests - One or more Spot Instance requests.

$sel:httpStatus:RequestSpotInstancesResponse', requestSpotInstancesResponse_httpStatus - The response's http status code.

Response Lenses

requestSpotInstancesResponse_spotInstanceRequests :: Lens' RequestSpotInstancesResponse (Maybe [SpotInstanceRequest])

One or more Spot Instance requests.

requestSpotInstancesResponse_httpStatus :: Lens' RequestSpotInstancesResponse Int

The response's http status code.