Network.AWS.EC2.ResetInstanceAttribute

Description

Resets an attribute of an instance to its default value. To reset the kernel or ramdisk, the instance must be in a stopped state. To reset the sourceDestCheck, the instance can be either running or stopped.

The sourceDestCheck attribute controls whether source/destination checking is enabled. The default value is true, which means checking is enabled. This value must be false for a NAT instance to perform NAT. For more information, see NAT Instances in the Amazon VPC User Guide.

Creating a Request

data ResetInstanceAttribute

See: newResetInstanceAttribute smart constructor.

Constructors

ResetInstanceAttribute'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

attribute :: InstanceAttributeName

The attribute to reset.

You can only reset the following attributes: kernel | ramdisk | sourceDestCheck. To change an instance attribute, use ModifyInstanceAttribute.

instanceId :: Text

The ID of the instance.

newResetInstanceAttribute

Arguments

:: InstanceAttributeName
$sel:attribute:ResetInstanceAttribute'

-> Text
$sel:instanceId:ResetInstanceAttribute'

-> ResetInstanceAttribute

Create a value of ResetInstanceAttribute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:ResetInstanceAttribute', resetInstanceAttribute_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:attribute:ResetInstanceAttribute', resetInstanceAttribute_attribute - The attribute to reset.

You can only reset the following attributes: kernel | ramdisk | sourceDestCheck. To change an instance attribute, use ModifyInstanceAttribute.

$sel:instanceId:ResetInstanceAttribute', resetInstanceAttribute_instanceId - The ID of the instance.

Request Lenses

resetInstanceAttribute_dryRun :: Lens' ResetInstanceAttribute (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

resetInstanceAttribute_attribute :: Lens' ResetInstanceAttribute InstanceAttributeName

The attribute to reset.

You can only reset the following attributes: kernel | ramdisk | sourceDestCheck. To change an instance attribute, use ModifyInstanceAttribute.

resetInstanceAttribute_instanceId :: Lens' ResetInstanceAttribute Text

The ID of the instance.

Destructuring the Response

data ResetInstanceAttributeResponse

See: newResetInstanceAttributeResponse smart constructor.

Constructors

ResetInstanceAttributeResponse'

newResetInstanceAttributeResponse :: ResetInstanceAttributeResponse

Create a value of ResetInstanceAttributeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.