Network.AWS.EC2.RestoreManagedPrefixListVersion

Description

Restores the entries from a previous version of a managed prefix list to a new version of the prefix list.

Creating a Request

data RestoreManagedPrefixListVersion

See: newRestoreManagedPrefixListVersion smart constructor.

Constructors

RestoreManagedPrefixListVersion'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

prefixListId :: Text

The ID of the prefix list.

previousVersion :: Integer

The version to restore.

currentVersion :: Integer

The current version number for the prefix list.

newRestoreManagedPrefixListVersion

Arguments

:: Text
$sel:prefixListId:RestoreManagedPrefixListVersion'

-> Integer
$sel:previousVersion:RestoreManagedPrefixListVersion'

-> Integer
$sel:currentVersion:RestoreManagedPrefixListVersion'

-> RestoreManagedPrefixListVersion

Create a value of RestoreManagedPrefixListVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:RestoreManagedPrefixListVersion', restoreManagedPrefixListVersion_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:prefixListId:RestoreManagedPrefixListVersion', restoreManagedPrefixListVersion_prefixListId - The ID of the prefix list.

$sel:previousVersion:RestoreManagedPrefixListVersion', restoreManagedPrefixListVersion_previousVersion - The version to restore.

$sel:currentVersion:RestoreManagedPrefixListVersion', restoreManagedPrefixListVersion_currentVersion - The current version number for the prefix list.

Request Lenses

restoreManagedPrefixListVersion_dryRun :: Lens' RestoreManagedPrefixListVersion (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

restoreManagedPrefixListVersion_prefixListId :: Lens' RestoreManagedPrefixListVersion Text

The ID of the prefix list.

restoreManagedPrefixListVersion_previousVersion :: Lens' RestoreManagedPrefixListVersion Integer

The version to restore.

restoreManagedPrefixListVersion_currentVersion :: Lens' RestoreManagedPrefixListVersion Integer

The current version number for the prefix list.

Destructuring the Response

data RestoreManagedPrefixListVersionResponse

See: newRestoreManagedPrefixListVersionResponse smart constructor.

Constructors

RestoreManagedPrefixListVersionResponse'

Fields

prefixList :: Maybe ManagedPrefixList

Information about the prefix list.

httpStatus :: Int

The response's http status code.

newRestoreManagedPrefixListVersionResponse

Arguments

:: Int
$sel:httpStatus:RestoreManagedPrefixListVersionResponse'

-> RestoreManagedPrefixListVersionResponse

Create a value of RestoreManagedPrefixListVersionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:prefixList:RestoreManagedPrefixListVersionResponse', restoreManagedPrefixListVersionResponse_prefixList - Information about the prefix list.

$sel:httpStatus:RestoreManagedPrefixListVersionResponse', restoreManagedPrefixListVersionResponse_httpStatus - The response's http status code.

Response Lenses

restoreManagedPrefixListVersionResponse_prefixList :: Lens' RestoreManagedPrefixListVersionResponse (Maybe ManagedPrefixList)

Information about the prefix list.

restoreManagedPrefixListVersionResponse_httpStatus :: Lens' RestoreManagedPrefixListVersionResponse Int

The response's http status code.