Network.AWS.EC2.RevokeClientVpnIngress

Description

Removes an ingress authorization rule from a Client VPN endpoint.

Creating a Request

data RevokeClientVpnIngress

See: newRevokeClientVpnIngress smart constructor.

Constructors

RevokeClientVpnIngress'

Fields

accessGroupId :: Maybe Text

The ID of the Active Directory group for which to revoke access.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

revokeAllGroups :: Maybe Bool

Indicates whether access should be revoked for all clients.

clientVpnEndpointId :: Text

The ID of the Client VPN endpoint with which the authorization rule is associated.

targetNetworkCidr :: Text

The IPv4 address range, in CIDR notation, of the network for which access is being removed.

newRevokeClientVpnIngress

Arguments

:: Text
$sel:clientVpnEndpointId:RevokeClientVpnIngress'

-> Text
$sel:targetNetworkCidr:RevokeClientVpnIngress'

-> RevokeClientVpnIngress

Create a value of RevokeClientVpnIngress with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accessGroupId:RevokeClientVpnIngress', revokeClientVpnIngress_accessGroupId - The ID of the Active Directory group for which to revoke access.

$sel:dryRun:RevokeClientVpnIngress', revokeClientVpnIngress_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:revokeAllGroups:RevokeClientVpnIngress', revokeClientVpnIngress_revokeAllGroups - Indicates whether access should be revoked for all clients.

$sel:clientVpnEndpointId:RevokeClientVpnIngress', revokeClientVpnIngress_clientVpnEndpointId - The ID of the Client VPN endpoint with which the authorization rule is associated.

$sel:targetNetworkCidr:RevokeClientVpnIngress', revokeClientVpnIngress_targetNetworkCidr - The IPv4 address range, in CIDR notation, of the network for which access is being removed.

Request Lenses

revokeClientVpnIngress_accessGroupId :: Lens' RevokeClientVpnIngress (Maybe Text)

The ID of the Active Directory group for which to revoke access.

revokeClientVpnIngress_dryRun :: Lens' RevokeClientVpnIngress (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

revokeClientVpnIngress_revokeAllGroups :: Lens' RevokeClientVpnIngress (Maybe Bool)

Indicates whether access should be revoked for all clients.

revokeClientVpnIngress_clientVpnEndpointId :: Lens' RevokeClientVpnIngress Text

The ID of the Client VPN endpoint with which the authorization rule is associated.

revokeClientVpnIngress_targetNetworkCidr :: Lens' RevokeClientVpnIngress Text

The IPv4 address range, in CIDR notation, of the network for which access is being removed.

Destructuring the Response

data RevokeClientVpnIngressResponse

See: newRevokeClientVpnIngressResponse smart constructor.

Constructors

RevokeClientVpnIngressResponse'

Fields

status :: Maybe ClientVpnAuthorizationRuleStatus

The current state of the authorization rule.

httpStatus :: Int

The response's http status code.

newRevokeClientVpnIngressResponse

Arguments

:: Int
$sel:httpStatus:RevokeClientVpnIngressResponse'

-> RevokeClientVpnIngressResponse

Create a value of RevokeClientVpnIngressResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:RevokeClientVpnIngressResponse', revokeClientVpnIngressResponse_status - The current state of the authorization rule.

$sel:httpStatus:RevokeClientVpnIngressResponse', revokeClientVpnIngressResponse_httpStatus - The response's http status code.

Response Lenses

revokeClientVpnIngressResponse_status :: Lens' RevokeClientVpnIngressResponse (Maybe ClientVpnAuthorizationRuleStatus)

The current state of the authorization rule.

revokeClientVpnIngressResponse_httpStatus :: Lens' RevokeClientVpnIngressResponse Int

The response's http status code.