Network.AWS.EC2.RevokeSecurityGroupEgress

Description

VPC only

Removes the specified outbound (egress) rules from a security group for EC2-VPC. This action does not apply to security groups for use in EC2-Classic.

You can specify rules using either rule IDs or security group rule properties. If you use rule properties, the values that you specify (for example, ports) must match the existing rule's values exactly. Each rule has a protocol, from and to ports, and destination (CIDR range, security group, or prefix list). For the TCP and UDP protocols, you must also specify the destination port or range of ports. For the ICMP protocol, you must also specify the ICMP type and code. If the security group rule has a description, you do not need to specify the description to revoke the rule.

Default VPC

If the values you specify do not match the existing rule's values, no error is returned, and the output describes the security group rules that were not revoked.

Amazon Web Services recommends that you describe the security group to verify that the rules were removed.

Rule changes are propagated to instances within the security group as quickly as possible. However, a small delay might occur.

Creating a Request

data RevokeSecurityGroupEgress

See: newRevokeSecurityGroupEgress smart constructor.

Constructors

RevokeSecurityGroupEgress'

Fields

fromPort :: Maybe Int

Not supported. Use a set of IP permissions to specify the port.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

sourceSecurityGroupName :: Maybe Text

Not supported. Use a set of IP permissions to specify a destination security group.

cidrIp :: Maybe Text

Not supported. Use a set of IP permissions to specify the CIDR.

ipProtocol :: Maybe Text

Not supported. Use a set of IP permissions to specify the protocol name or number.

ipPermissions :: Maybe [IpPermission]

The sets of IP permissions. You can't specify a destination security group and a CIDR IP address range in the same set of permissions.

securityGroupRuleIds :: Maybe [Text]

The IDs of the security group rules.

sourceSecurityGroupOwnerId :: Maybe Text

Not supported. Use a set of IP permissions to specify a destination security group.

toPort :: Maybe Int

Not supported. Use a set of IP permissions to specify the port.

groupId :: Text

The ID of the security group.

newRevokeSecurityGroupEgress

Arguments

:: Text
$sel:groupId:RevokeSecurityGroupEgress'

-> RevokeSecurityGroupEgress

Create a value of RevokeSecurityGroupEgress with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fromPort:RevokeSecurityGroupEgress', revokeSecurityGroupEgress_fromPort - Not supported. Use a set of IP permissions to specify the port.

$sel:dryRun:RevokeSecurityGroupEgress', revokeSecurityGroupEgress_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:sourceSecurityGroupName:RevokeSecurityGroupEgress', revokeSecurityGroupEgress_sourceSecurityGroupName - Not supported. Use a set of IP permissions to specify a destination security group.

$sel:cidrIp:RevokeSecurityGroupEgress', revokeSecurityGroupEgress_cidrIp - Not supported. Use a set of IP permissions to specify the CIDR.

$sel:ipProtocol:RevokeSecurityGroupEgress', revokeSecurityGroupEgress_ipProtocol - Not supported. Use a set of IP permissions to specify the protocol name or number.

$sel:ipPermissions:RevokeSecurityGroupEgress', revokeSecurityGroupEgress_ipPermissions - The sets of IP permissions. You can't specify a destination security group and a CIDR IP address range in the same set of permissions.

$sel:securityGroupRuleIds:RevokeSecurityGroupEgress', revokeSecurityGroupEgress_securityGroupRuleIds - The IDs of the security group rules.

$sel:sourceSecurityGroupOwnerId:RevokeSecurityGroupEgress', revokeSecurityGroupEgress_sourceSecurityGroupOwnerId - Not supported. Use a set of IP permissions to specify a destination security group.

$sel:toPort:RevokeSecurityGroupEgress', revokeSecurityGroupEgress_toPort - Not supported. Use a set of IP permissions to specify the port.

$sel:groupId:RevokeSecurityGroupEgress', revokeSecurityGroupEgress_groupId - The ID of the security group.

Request Lenses

revokeSecurityGroupEgress_fromPort :: Lens' RevokeSecurityGroupEgress (Maybe Int)

Not supported. Use a set of IP permissions to specify the port.

revokeSecurityGroupEgress_dryRun :: Lens' RevokeSecurityGroupEgress (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

revokeSecurityGroupEgress_sourceSecurityGroupName :: Lens' RevokeSecurityGroupEgress (Maybe Text)

Not supported. Use a set of IP permissions to specify a destination security group.

revokeSecurityGroupEgress_cidrIp :: Lens' RevokeSecurityGroupEgress (Maybe Text)

Not supported. Use a set of IP permissions to specify the CIDR.

revokeSecurityGroupEgress_ipProtocol :: Lens' RevokeSecurityGroupEgress (Maybe Text)

Not supported. Use a set of IP permissions to specify the protocol name or number.

revokeSecurityGroupEgress_ipPermissions :: Lens' RevokeSecurityGroupEgress (Maybe [IpPermission])

The sets of IP permissions. You can't specify a destination security group and a CIDR IP address range in the same set of permissions.

revokeSecurityGroupEgress_securityGroupRuleIds :: Lens' RevokeSecurityGroupEgress (Maybe [Text])

The IDs of the security group rules.

revokeSecurityGroupEgress_sourceSecurityGroupOwnerId :: Lens' RevokeSecurityGroupEgress (Maybe Text)

Not supported. Use a set of IP permissions to specify a destination security group.

revokeSecurityGroupEgress_toPort :: Lens' RevokeSecurityGroupEgress (Maybe Int)

Not supported. Use a set of IP permissions to specify the port.

revokeSecurityGroupEgress_groupId :: Lens' RevokeSecurityGroupEgress Text

The ID of the security group.

Destructuring the Response

data RevokeSecurityGroupEgressResponse

See: newRevokeSecurityGroupEgressResponse smart constructor.

Constructors

RevokeSecurityGroupEgressResponse'

Fields

unknownIpPermissions :: Maybe [IpPermission]

The outbound rules that were unknown to the service. In some cases, unknownIpPermissionSet might be in a different format from the request parameter.

return' :: Maybe Bool

Returns true if the request succeeds; otherwise, returns an error.

httpStatus :: Int

The response's http status code.

newRevokeSecurityGroupEgressResponse

Arguments

:: Int
$sel:httpStatus:RevokeSecurityGroupEgressResponse'

-> RevokeSecurityGroupEgressResponse

Create a value of RevokeSecurityGroupEgressResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:unknownIpPermissions:RevokeSecurityGroupEgressResponse', revokeSecurityGroupEgressResponse_unknownIpPermissions - The outbound rules that were unknown to the service. In some cases, unknownIpPermissionSet might be in a different format from the request parameter.

$sel:return':RevokeSecurityGroupEgressResponse', revokeSecurityGroupEgressResponse_return - Returns true if the request succeeds; otherwise, returns an error.

$sel:httpStatus:RevokeSecurityGroupEgressResponse', revokeSecurityGroupEgressResponse_httpStatus - The response's http status code.

Response Lenses

revokeSecurityGroupEgressResponse_unknownIpPermissions :: Lens' RevokeSecurityGroupEgressResponse (Maybe [IpPermission])

The outbound rules that were unknown to the service. In some cases, unknownIpPermissionSet might be in a different format from the request parameter.

revokeSecurityGroupEgressResponse_return :: Lens' RevokeSecurityGroupEgressResponse (Maybe Bool)

Returns true if the request succeeds; otherwise, returns an error.

revokeSecurityGroupEgressResponse_httpStatus :: Lens' RevokeSecurityGroupEgressResponse Int

The response's http status code.