Network.AWS.EC2.RevokeSecurityGroupIngress

Description

Removes the specified inbound (ingress) rules from a security group.

You can specify rules using either rule IDs or security group rule properties. If you use rule properties, the values that you specify (for example, ports) must match the existing rule's values exactly. Each rule has a protocol, from and to ports, and source (CIDR range, security group, or prefix list). For the TCP and UDP protocols, you must also specify the destination port or range of ports. For the ICMP protocol, you must also specify the ICMP type and code. If the security group rule has a description, you do not need to specify the description to revoke the rule.

EC2-Classic, default VPC

If the values you specify do not match the existing rule's values, no error is returned, and the output describes the security group rules that were not revoked.

Amazon Web Services recommends that you describe the security group to verify that the rules were removed.

Rule changes are propagated to instances within the security group as quickly as possible. However, a small delay might occur.

Creating a Request

data RevokeSecurityGroupIngress

See: newRevokeSecurityGroupIngress smart constructor.

Constructors

RevokeSecurityGroupIngress'

Fields

fromPort :: Maybe Int

The start of port range for the TCP and UDP protocols, or an ICMP type number. For the ICMP type number, use -1 to specify all ICMP types.

ipPermissions :: Maybe [IpPermission]

The sets of IP permissions. You can't specify a source security group and a CIDR IP address range in the same set of permissions.

ipProtocol :: Maybe Text

The IP protocol name (tcp, udp, icmp) or number (see Protocol Numbers). Use -1 to specify all.

groupId :: Maybe Text

The ID of the security group. You must specify either the security group ID or the security group name in the request. For security groups in a nondefault VPC, you must specify the security group ID.

toPort :: Maybe Int

The end of port range for the TCP and UDP protocols, or an ICMP code number. For the ICMP code number, use -1 to specify all ICMP codes for the ICMP type.

cidrIp :: Maybe Text

The CIDR IP address range. You can't specify this parameter when specifying a source security group.

sourceSecurityGroupOwnerId :: Maybe Text

EC2-Classic

The Amazon Web Services account ID of the source security group, if the source security group is in a different account. You can't specify this parameter in combination with the following parameters: the CIDR IP address range, the IP protocol, the start of the port range, and the end of the port range. To revoke a specific rule for an IP protocol and port range, use a set of IP permissions instead.

groupName :: Maybe Text

EC2-Classic, default VPC

The name of the security group. You must specify either the security group ID or the security group name in the request.

sourceSecurityGroupName :: Maybe Text

EC2-Classic, default VPC

The name of the source security group. You can't specify this parameter in combination with the following parameters: the CIDR IP address range, the start of the port range, the IP protocol, and the end of the port range. For EC2-VPC, the source security group must be in the same VPC. To revoke a specific rule for an IP protocol and port range, use a set of IP permissions instead.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

securityGroupRuleIds :: Maybe [Text]

The IDs of the security group rules.

newRevokeSecurityGroupIngress :: RevokeSecurityGroupIngress

Create a value of RevokeSecurityGroupIngress with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fromPort:RevokeSecurityGroupIngress', revokeSecurityGroupIngress_fromPort - The start of port range for the TCP and UDP protocols, or an ICMP type number. For the ICMP type number, use -1 to specify all ICMP types.

$sel:ipPermissions:RevokeSecurityGroupIngress', revokeSecurityGroupIngress_ipPermissions - The sets of IP permissions. You can't specify a source security group and a CIDR IP address range in the same set of permissions.

$sel:ipProtocol:RevokeSecurityGroupIngress', revokeSecurityGroupIngress_ipProtocol - The IP protocol name (tcp, udp, icmp) or number (see Protocol Numbers). Use -1 to specify all.

$sel:groupId:RevokeSecurityGroupIngress', revokeSecurityGroupIngress_groupId - The ID of the security group. You must specify either the security group ID or the security group name in the request. For security groups in a nondefault VPC, you must specify the security group ID.

$sel:toPort:RevokeSecurityGroupIngress', revokeSecurityGroupIngress_toPort - The end of port range for the TCP and UDP protocols, or an ICMP code number. For the ICMP code number, use -1 to specify all ICMP codes for the ICMP type.

$sel:cidrIp:RevokeSecurityGroupIngress', revokeSecurityGroupIngress_cidrIp - The CIDR IP address range. You can't specify this parameter when specifying a source security group.

$sel:sourceSecurityGroupOwnerId:RevokeSecurityGroupIngress', revokeSecurityGroupIngress_sourceSecurityGroupOwnerId - [EC2-Classic] The Amazon Web Services account ID of the source security group, if the source security group is in a different account. You can't specify this parameter in combination with the following parameters: the CIDR IP address range, the IP protocol, the start of the port range, and the end of the port range. To revoke a specific rule for an IP protocol and port range, use a set of IP permissions instead.

$sel:groupName:RevokeSecurityGroupIngress', revokeSecurityGroupIngress_groupName - [EC2-Classic, default VPC] The name of the security group. You must specify either the security group ID or the security group name in the request.

$sel:sourceSecurityGroupName:RevokeSecurityGroupIngress', revokeSecurityGroupIngress_sourceSecurityGroupName - [EC2-Classic, default VPC] The name of the source security group. You can't specify this parameter in combination with the following parameters: the CIDR IP address range, the start of the port range, the IP protocol, and the end of the port range. For EC2-VPC, the source security group must be in the same VPC. To revoke a specific rule for an IP protocol and port range, use a set of IP permissions instead.

$sel:dryRun:RevokeSecurityGroupIngress', revokeSecurityGroupIngress_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:securityGroupRuleIds:RevokeSecurityGroupIngress', revokeSecurityGroupIngress_securityGroupRuleIds - The IDs of the security group rules.

Request Lenses

revokeSecurityGroupIngress_fromPort :: Lens' RevokeSecurityGroupIngress (Maybe Int)

The start of port range for the TCP and UDP protocols, or an ICMP type number. For the ICMP type number, use -1 to specify all ICMP types.

revokeSecurityGroupIngress_ipPermissions :: Lens' RevokeSecurityGroupIngress (Maybe [IpPermission])

The sets of IP permissions. You can't specify a source security group and a CIDR IP address range in the same set of permissions.

revokeSecurityGroupIngress_ipProtocol :: Lens' RevokeSecurityGroupIngress (Maybe Text)

The IP protocol name (tcp, udp, icmp) or number (see Protocol Numbers). Use -1 to specify all.

revokeSecurityGroupIngress_groupId :: Lens' RevokeSecurityGroupIngress (Maybe Text)

The ID of the security group. You must specify either the security group ID or the security group name in the request. For security groups in a nondefault VPC, you must specify the security group ID.

revokeSecurityGroupIngress_toPort :: Lens' RevokeSecurityGroupIngress (Maybe Int)

The end of port range for the TCP and UDP protocols, or an ICMP code number. For the ICMP code number, use -1 to specify all ICMP codes for the ICMP type.

revokeSecurityGroupIngress_cidrIp :: Lens' RevokeSecurityGroupIngress (Maybe Text)

The CIDR IP address range. You can't specify this parameter when specifying a source security group.

revokeSecurityGroupIngress_sourceSecurityGroupOwnerId :: Lens' RevokeSecurityGroupIngress (Maybe Text)

EC2-Classic

The Amazon Web Services account ID of the source security group, if the source security group is in a different account. You can't specify this parameter in combination with the following parameters: the CIDR IP address range, the IP protocol, the start of the port range, and the end of the port range. To revoke a specific rule for an IP protocol and port range, use a set of IP permissions instead.

revokeSecurityGroupIngress_groupName :: Lens' RevokeSecurityGroupIngress (Maybe Text)

EC2-Classic, default VPC

The name of the security group. You must specify either the security group ID or the security group name in the request.

revokeSecurityGroupIngress_sourceSecurityGroupName :: Lens' RevokeSecurityGroupIngress (Maybe Text)

EC2-Classic, default VPC

The name of the source security group. You can't specify this parameter in combination with the following parameters: the CIDR IP address range, the start of the port range, the IP protocol, and the end of the port range. For EC2-VPC, the source security group must be in the same VPC. To revoke a specific rule for an IP protocol and port range, use a set of IP permissions instead.

revokeSecurityGroupIngress_dryRun :: Lens' RevokeSecurityGroupIngress (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

revokeSecurityGroupIngress_securityGroupRuleIds :: Lens' RevokeSecurityGroupIngress (Maybe [Text])

The IDs of the security group rules.

Destructuring the Response

data RevokeSecurityGroupIngressResponse

See: newRevokeSecurityGroupIngressResponse smart constructor.

Constructors

RevokeSecurityGroupIngressResponse'

Fields

return' :: Maybe Bool

Returns true if the request succeeds; otherwise, returns an error.

unknownIpPermissions :: Maybe [IpPermission]

The inbound rules that were unknown to the service. In some cases, unknownIpPermissionSet might be in a different format from the request parameter.

httpStatus :: Int

The response's http status code.

newRevokeSecurityGroupIngressResponse

Arguments

:: Int
$sel:httpStatus:RevokeSecurityGroupIngressResponse'

-> RevokeSecurityGroupIngressResponse

Create a value of RevokeSecurityGroupIngressResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:return':RevokeSecurityGroupIngressResponse', revokeSecurityGroupIngressResponse_return - Returns true if the request succeeds; otherwise, returns an error.

$sel:unknownIpPermissions:RevokeSecurityGroupIngressResponse', revokeSecurityGroupIngressResponse_unknownIpPermissions - The inbound rules that were unknown to the service. In some cases, unknownIpPermissionSet might be in a different format from the request parameter.

$sel:httpStatus:RevokeSecurityGroupIngressResponse', revokeSecurityGroupIngressResponse_httpStatus - The response's http status code.

Response Lenses

revokeSecurityGroupIngressResponse_return :: Lens' RevokeSecurityGroupIngressResponse (Maybe Bool)

Returns true if the request succeeds; otherwise, returns an error.

revokeSecurityGroupIngressResponse_unknownIpPermissions :: Lens' RevokeSecurityGroupIngressResponse (Maybe [IpPermission])

The inbound rules that were unknown to the service. In some cases, unknownIpPermissionSet might be in a different format from the request parameter.

revokeSecurityGroupIngressResponse_httpStatus :: Lens' RevokeSecurityGroupIngressResponse Int

The response's http status code.