Network.AWS.EC2.RunInstances

Description

Launches the specified number of instances using an AMI for which you have permissions.

You can specify a number of options, or leave the default options. The following rules apply:

[EC2-VPC] If you don't specify a subnet ID, we choose a default subnet from your default VPC for you. If you don't have a default VPC, you must specify a subnet ID in the request.

[EC2-Classic] If don't specify an Availability Zone, we choose one for you.

Some instance types must be launched into a VPC. If you do not have a default VPC, or if you do not specify a subnet ID, the request fails. For more information, see Instance types available only in a VPC.

[EC2-VPC] All instances have a network interface with a primary private IPv4 address. If you don't specify this address, we choose one from the IPv4 range of your subnet.

Not all instance types support IPv6 addresses. For more information, see Instance types.

If you don't specify a security group ID, we use the default security group. For more information, see Security groups.

If any of the AMIs have a product code attached for which the user has not subscribed, the request fails.

You can create a launch template, which is a resource that contains the parameters to launch an instance. When you launch an instance using RunInstances, you can specify the launch template instead of specifying the launch parameters.

To ensure faster instance launches, break up large requests into smaller batches. For example, create five separate launch requests for 100 instances each instead of one launch request for 500 instances.

An instance is ready for you to use when it's in the running state. You can check the state of your instance using DescribeInstances. You can tag instances and EBS volumes during launch, after launch, or both. For more information, see CreateTags and Tagging your Amazon EC2 resources.

Linux instances have access to the public key of the key pair at boot. You can use this key to provide secure access to the instance. Amazon EC2 public images use this feature to provide secure access without passwords. For more information, see Key pairs.

For troubleshooting, see What to do if an instance immediately terminates, and Troubleshooting connecting to your instance.

Creating a Request

data RunInstances

See: newRunInstances smart constructor.

Constructors

RunInstances'

Fields

additionalInfo :: Maybe Text

Reserved.

securityGroupIds :: Maybe [Text]

The IDs of the security groups. You can create a security group using CreateSecurityGroup.

If you specify a network interface, you must specify any security groups as part of the network interface.

tagSpecifications :: Maybe [TagSpecification]

The tags to apply to the resources during launch. You can only tag instances and volumes on launch. The specified tags are applied to all instances or volumes that are created during launch. To tag a resource after it has been created, see CreateTags.

capacityReservationSpecification :: Maybe CapacityReservationSpecification

Information about the Capacity Reservation targeting option. If you do not specify this parameter, the instance's Capacity Reservation preference defaults to open, which enables it to run in any open Capacity Reservation that has matching attributes (instance type, platform, Availability Zone).

ebsOptimized :: Maybe Bool

Indicates whether the instance is optimized for Amazon EBS I/O. This optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal Amazon EBS I/O performance. This optimization isn't available with all instance types. Additional usage charges apply when using an EBS-optimized instance.

Default: false

placement :: Maybe Placement

The placement for the instance.

userData :: Maybe Text

The user data to make available to the instance. For more information, see Running commands on your Linux instance at launch (Linux) and Adding User Data (Windows). If you are using a command line tool, base64-encoding is performed for you, and you can load the text from a file. Otherwise, you must provide base64-encoded text. User data is limited to 16 KB.

instanceType :: Maybe InstanceType

The instance type. For more information, see Instance types in the Amazon EC2 User Guide.

Default: m1.small

ipv6Addresses :: Maybe [InstanceIpv6Address]

EC2-VPC

The IPv6 addresses from the range of the subnet to associate with the primary network interface. You cannot specify this option and the option to assign a number of IPv6 addresses in the same request. You cannot specify this option if you've specified a minimum number of instances to launch.

You cannot specify this option and the network interfaces option in the same request.

ramdiskId :: Maybe Text

The ID of the RAM disk to select. Some kernels require additional drivers at launch. Check the kernel requirements for information about whether you need to specify a RAM disk. To find kernel requirements, go to the Amazon Web Services Resource Center and search for the kernel ID.

We recommend that you use PV-GRUB instead of kernels and RAM disks. For more information, see PV-GRUB in the Amazon EC2 User Guide.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

creditSpecification :: Maybe CreditSpecificationRequest

The credit option for CPU usage of the burstable performance instance. Valid values are standard and unlimited. To change this attribute after launch, use ModifyInstanceCreditSpecification. For more information, see Burstable performance instances in the Amazon EC2 User Guide.

Default: standard (T2 instances) or unlimited (T3/T3a instances)

For T3 instances with host tenancy, only standard is supported.

licenseSpecifications :: Maybe [LicenseConfigurationRequest]

The license configurations.

instanceMarketOptions :: Maybe InstanceMarketOptionsRequest

The market (purchasing) option for the instances.

For RunInstances, persistent Spot Instance requests are only supported when InstanceInterruptionBehavior is set to either hibernate or stop.

launchTemplate :: Maybe LaunchTemplateSpecification

The launch template to use to launch the instances. Any parameters that you specify in RunInstances override the same parameters in the launch template. You can specify either the name or ID of a launch template, but not both.

instanceInitiatedShutdownBehavior :: Maybe ShutdownBehavior

Indicates whether an instance stops or terminates when you initiate shutdown from the instance (using the operating system command for system shutdown).

Default: stop

elasticInferenceAccelerators :: Maybe [ElasticInferenceAccelerator]

An elastic inference accelerator to associate with the instance. Elastic inference accelerators are a resource you can attach to your Amazon EC2 instances to accelerate your Deep Learning (DL) inference workloads.

You cannot specify accelerators from different generations in the same request.

imageId :: Maybe Text

The ID of the AMI. An AMI ID is required to launch an instance and must be specified here or in a launch template.

securityGroups :: Maybe [Text]

EC2-Classic, default VPC

The names of the security groups. For a nondefault VPC, you must use security group IDs instead.

If you specify a network interface, you must specify any security groups as part of the network interface.

Default: Amazon EC2 uses the default security group.

elasticGpuSpecification :: Maybe [ElasticGpuSpecification]

An elastic GPU to associate with the instance. An Elastic GPU is a GPU resource that you can attach to your Windows instance to accelerate the graphics performance of your applications. For more information, see Amazon EC2 Elastic GPUs in the Amazon EC2 User Guide.

iamInstanceProfile :: Maybe IamInstanceProfileSpecification

The name or Amazon Resource Name (ARN) of an IAM instance profile.

hibernationOptions :: Maybe HibernationOptionsRequest

Indicates whether an instance is enabled for hibernation. For more information, see Hibernate your instance in the Amazon EC2 User Guide.

You can't enable hibernation and Amazon Web Services Nitro Enclaves on the same instance.

ipv6AddressCount :: Maybe Int

EC2-VPC

The number of IPv6 addresses to associate with the primary network interface. Amazon EC2 chooses the IPv6 addresses from the range of your subnet. You cannot specify this option and the option to assign specific IPv6 addresses in the same request. You can specify this option if you've specified a minimum number of instances to launch.

You cannot specify this option and the network interfaces option in the same request.

monitoring :: Maybe RunInstancesMonitoringEnabled

Specifies whether detailed monitoring is enabled for the instance.

blockDeviceMappings :: Maybe [BlockDeviceMapping]

The block device mapping, which defines the EBS volumes and instance store volumes to attach to the instance at launch. For more information, see Block device mappings in the Amazon EC2 User Guide.

subnetId :: Maybe Text

EC2-VPC

The ID of the subnet to launch the instance into.

If you specify a network interface, you must specify any subnets as part of the network interface.

enclaveOptions :: Maybe EnclaveOptionsRequest

Indicates whether the instance is enabled for Amazon Web Services Nitro Enclaves. For more information, see What is Amazon Web Services Nitro Enclaves? in the Amazon Web Services Nitro Enclaves User Guide.

You can't enable Amazon Web Services Nitro Enclaves and hibernation on the same instance.

kernelId :: Maybe Text

The ID of the kernel.

We recommend that you use PV-GRUB instead of kernels and RAM disks. For more information, see PV-GRUB in the Amazon EC2 User Guide.

cpuOptions :: Maybe CpuOptionsRequest

The CPU options for the instance. For more information, see Optimizing CPU options in the Amazon EC2 User Guide.

disableApiTermination :: Maybe Bool

If you set this parameter to true, you can't terminate the instance using the Amazon EC2 console, CLI, or API; otherwise, you can. To change this attribute after launch, use ModifyInstanceAttribute. Alternatively, if you set InstanceInitiatedShutdownBehavior to terminate, you can terminate the instance by running the shutdown command from the instance.

Default: false

networkInterfaces :: Maybe [InstanceNetworkInterfaceSpecification]

The network interfaces to associate with the instance. If you specify a network interface, you must specify any security groups and subnets as part of the network interface.

keyName :: Maybe Text

The name of the key pair. You can create a key pair using CreateKeyPair or ImportKeyPair.

If you do not specify a key pair, you can't connect to the instance unless you choose an AMI that is configured to allow users another way to log in.

metadataOptions :: Maybe InstanceMetadataOptionsRequest

The metadata options for the instance. For more information, see Instance metadata and user data.

privateIpAddress :: Maybe Text

EC2-VPC

The primary IPv4 address. You must specify a value from the IPv4 address range of the subnet.

Only one private IP address can be designated as primary. You can't specify this option if you've specified the option to designate a private IP address as the primary IP address in a network interface specification. You cannot specify this option if you're launching more than one instance in the request.

You cannot specify this option and the network interfaces option in the same request.

clientToken :: Maybe Text

Unique, case-sensitive identifier you provide to ensure the idempotency of the request. If you do not specify a client token, a randomly generated token is used for the request to ensure idempotency.

For more information, see Ensuring Idempotency.

Constraints: Maximum 64 ASCII characters

maxCount :: Int

The maximum number of instances to launch. If you specify more instances than Amazon EC2 can launch in the target Availability Zone, Amazon EC2 launches the largest possible number of instances above MinCount.

Constraints: Between 1 and the maximum number you're allowed for the specified instance type. For more information about the default limits, and how to request an increase, see How many instances can I run in Amazon EC2 in the Amazon EC2 FAQ.

minCount :: Int

The minimum number of instances to launch. If you specify a minimum that is more instances than Amazon EC2 can launch in the target Availability Zone, Amazon EC2 launches no instances.

Constraints: Between 1 and the maximum number you're allowed for the specified instance type. For more information about the default limits, and how to request an increase, see How many instances can I run in Amazon EC2 in the Amazon EC2 General FAQ.

newRunInstances

Arguments

:: Int
$sel:maxCount:RunInstances'

-> Int
$sel:minCount:RunInstances'

-> RunInstances

Create a value of RunInstances with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:additionalInfo:RunInstances', runInstances_additionalInfo - Reserved.

$sel:securityGroupIds:RunInstances', runInstances_securityGroupIds - The IDs of the security groups. You can create a security group using CreateSecurityGroup.

If you specify a network interface, you must specify any security groups as part of the network interface.

$sel:tagSpecifications:RunInstances', runInstances_tagSpecifications - The tags to apply to the resources during launch. You can only tag instances and volumes on launch. The specified tags are applied to all instances or volumes that are created during launch. To tag a resource after it has been created, see CreateTags.

$sel:capacityReservationSpecification:RunInstances', runInstances_capacityReservationSpecification - Information about the Capacity Reservation targeting option. If you do not specify this parameter, the instance's Capacity Reservation preference defaults to open, which enables it to run in any open Capacity Reservation that has matching attributes (instance type, platform, Availability Zone).

$sel:ebsOptimized:RunInstances', runInstances_ebsOptimized - Indicates whether the instance is optimized for Amazon EBS I/O. This optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal Amazon EBS I/O performance. This optimization isn't available with all instance types. Additional usage charges apply when using an EBS-optimized instance.

Default: false

$sel:placement:RunInstances', runInstances_placement - The placement for the instance.

$sel:userData:RunInstances', runInstances_userData - The user data to make available to the instance. For more information, see Running commands on your Linux instance at launch (Linux) and Adding User Data (Windows). If you are using a command line tool, base64-encoding is performed for you, and you can load the text from a file. Otherwise, you must provide base64-encoded text. User data is limited to 16 KB.

$sel:instanceType:RunInstances', runInstances_instanceType - The instance type. For more information, see Instance types in the Amazon EC2 User Guide.

Default: m1.small

$sel:ipv6Addresses:RunInstances', runInstances_ipv6Addresses - [EC2-VPC] The IPv6 addresses from the range of the subnet to associate with the primary network interface. You cannot specify this option and the option to assign a number of IPv6 addresses in the same request. You cannot specify this option if you've specified a minimum number of instances to launch.

You cannot specify this option and the network interfaces option in the same request.

$sel:ramdiskId:RunInstances', runInstances_ramdiskId - The ID of the RAM disk to select. Some kernels require additional drivers at launch. Check the kernel requirements for information about whether you need to specify a RAM disk. To find kernel requirements, go to the Amazon Web Services Resource Center and search for the kernel ID.

We recommend that you use PV-GRUB instead of kernels and RAM disks. For more information, see PV-GRUB in the Amazon EC2 User Guide.

$sel:dryRun:RunInstances', runInstances_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:creditSpecification:RunInstances', runInstances_creditSpecification - The credit option for CPU usage of the burstable performance instance. Valid values are standard and unlimited. To change this attribute after launch, use ModifyInstanceCreditSpecification. For more information, see Burstable performance instances in the Amazon EC2 User Guide.

Default: standard (T2 instances) or unlimited (T3/T3a instances)

For T3 instances with host tenancy, only standard is supported.

$sel:licenseSpecifications:RunInstances', runInstances_licenseSpecifications - The license configurations.

$sel:instanceMarketOptions:RunInstances', runInstances_instanceMarketOptions - The market (purchasing) option for the instances.

For RunInstances, persistent Spot Instance requests are only supported when InstanceInterruptionBehavior is set to either hibernate or stop.

$sel:launchTemplate:RunInstances', runInstances_launchTemplate - The launch template to use to launch the instances. Any parameters that you specify in RunInstances override the same parameters in the launch template. You can specify either the name or ID of a launch template, but not both.

$sel:instanceInitiatedShutdownBehavior:RunInstances', runInstances_instanceInitiatedShutdownBehavior - Indicates whether an instance stops or terminates when you initiate shutdown from the instance (using the operating system command for system shutdown).

Default: stop

$sel:elasticInferenceAccelerators:RunInstances', runInstances_elasticInferenceAccelerators - An elastic inference accelerator to associate with the instance. Elastic inference accelerators are a resource you can attach to your Amazon EC2 instances to accelerate your Deep Learning (DL) inference workloads.

You cannot specify accelerators from different generations in the same request.

$sel:imageId:RunInstances', runInstances_imageId - The ID of the AMI. An AMI ID is required to launch an instance and must be specified here or in a launch template.

$sel:securityGroups:RunInstances', runInstances_securityGroups - [EC2-Classic, default VPC] The names of the security groups. For a nondefault VPC, you must use security group IDs instead.

If you specify a network interface, you must specify any security groups as part of the network interface.

Default: Amazon EC2 uses the default security group.

$sel:elasticGpuSpecification:RunInstances', runInstances_elasticGpuSpecification - An elastic GPU to associate with the instance. An Elastic GPU is a GPU resource that you can attach to your Windows instance to accelerate the graphics performance of your applications. For more information, see Amazon EC2 Elastic GPUs in the Amazon EC2 User Guide.

$sel:iamInstanceProfile:RunInstances', runInstances_iamInstanceProfile - The name or Amazon Resource Name (ARN) of an IAM instance profile.

$sel:hibernationOptions:RunInstances', runInstances_hibernationOptions - Indicates whether an instance is enabled for hibernation. For more information, see Hibernate your instance in the Amazon EC2 User Guide.

You can't enable hibernation and Amazon Web Services Nitro Enclaves on the same instance.

$sel:ipv6AddressCount:RunInstances', runInstances_ipv6AddressCount - [EC2-VPC] The number of IPv6 addresses to associate with the primary network interface. Amazon EC2 chooses the IPv6 addresses from the range of your subnet. You cannot specify this option and the option to assign specific IPv6 addresses in the same request. You can specify this option if you've specified a minimum number of instances to launch.

You cannot specify this option and the network interfaces option in the same request.

$sel:monitoring:RunInstances', runInstances_monitoring - Specifies whether detailed monitoring is enabled for the instance.

$sel:blockDeviceMappings:RunInstances', runInstances_blockDeviceMappings - The block device mapping, which defines the EBS volumes and instance store volumes to attach to the instance at launch. For more information, see Block device mappings in the Amazon EC2 User Guide.

$sel:subnetId:RunInstances', runInstances_subnetId - [EC2-VPC] The ID of the subnet to launch the instance into.

If you specify a network interface, you must specify any subnets as part of the network interface.

$sel:enclaveOptions:RunInstances', runInstances_enclaveOptions - Indicates whether the instance is enabled for Amazon Web Services Nitro Enclaves. For more information, see What is Amazon Web Services Nitro Enclaves? in the Amazon Web Services Nitro Enclaves User Guide.

You can't enable Amazon Web Services Nitro Enclaves and hibernation on the same instance.

$sel:kernelId:RunInstances', runInstances_kernelId - The ID of the kernel.

We recommend that you use PV-GRUB instead of kernels and RAM disks. For more information, see PV-GRUB in the Amazon EC2 User Guide.

$sel:cpuOptions:RunInstances', runInstances_cpuOptions - The CPU options for the instance. For more information, see Optimizing CPU options in the Amazon EC2 User Guide.

$sel:disableApiTermination:RunInstances', runInstances_disableApiTermination - If you set this parameter to true, you can't terminate the instance using the Amazon EC2 console, CLI, or API; otherwise, you can. To change this attribute after launch, use ModifyInstanceAttribute. Alternatively, if you set InstanceInitiatedShutdownBehavior to terminate, you can terminate the instance by running the shutdown command from the instance.

Default: false

$sel:networkInterfaces:RunInstances', runInstances_networkInterfaces - The network interfaces to associate with the instance. If you specify a network interface, you must specify any security groups and subnets as part of the network interface.

$sel:keyName:RunInstances', runInstances_keyName - The name of the key pair. You can create a key pair using CreateKeyPair or ImportKeyPair.

If you do not specify a key pair, you can't connect to the instance unless you choose an AMI that is configured to allow users another way to log in.

$sel:metadataOptions:RunInstances', runInstances_metadataOptions - The metadata options for the instance. For more information, see Instance metadata and user data.

$sel:privateIpAddress:RunInstances', runInstances_privateIpAddress - [EC2-VPC] The primary IPv4 address. You must specify a value from the IPv4 address range of the subnet.

Only one private IP address can be designated as primary. You can't specify this option if you've specified the option to designate a private IP address as the primary IP address in a network interface specification. You cannot specify this option if you're launching more than one instance in the request.

You cannot specify this option and the network interfaces option in the same request.

$sel:clientToken:RunInstances', runInstances_clientToken - Unique, case-sensitive identifier you provide to ensure the idempotency of the request. If you do not specify a client token, a randomly generated token is used for the request to ensure idempotency.

For more information, see Ensuring Idempotency.

Constraints: Maximum 64 ASCII characters

$sel:maxCount:RunInstances', runInstances_maxCount - The maximum number of instances to launch. If you specify more instances than Amazon EC2 can launch in the target Availability Zone, Amazon EC2 launches the largest possible number of instances above MinCount.

Constraints: Between 1 and the maximum number you're allowed for the specified instance type. For more information about the default limits, and how to request an increase, see How many instances can I run in Amazon EC2 in the Amazon EC2 FAQ.

$sel:minCount:RunInstances', runInstances_minCount - The minimum number of instances to launch. If you specify a minimum that is more instances than Amazon EC2 can launch in the target Availability Zone, Amazon EC2 launches no instances.

Constraints: Between 1 and the maximum number you're allowed for the specified instance type. For more information about the default limits, and how to request an increase, see How many instances can I run in Amazon EC2 in the Amazon EC2 General FAQ.

Request Lenses

runInstances_additionalInfo :: Lens' RunInstances (Maybe Text)

Reserved.

runInstances_securityGroupIds :: Lens' RunInstances (Maybe [Text])

The IDs of the security groups. You can create a security group using CreateSecurityGroup.

If you specify a network interface, you must specify any security groups as part of the network interface.

runInstances_tagSpecifications :: Lens' RunInstances (Maybe [TagSpecification])

The tags to apply to the resources during launch. You can only tag instances and volumes on launch. The specified tags are applied to all instances or volumes that are created during launch. To tag a resource after it has been created, see CreateTags.

runInstances_capacityReservationSpecification :: Lens' RunInstances (Maybe CapacityReservationSpecification)

Information about the Capacity Reservation targeting option. If you do not specify this parameter, the instance's Capacity Reservation preference defaults to open, which enables it to run in any open Capacity Reservation that has matching attributes (instance type, platform, Availability Zone).

runInstances_ebsOptimized :: Lens' RunInstances (Maybe Bool)

Indicates whether the instance is optimized for Amazon EBS I/O. This optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal Amazon EBS I/O performance. This optimization isn't available with all instance types. Additional usage charges apply when using an EBS-optimized instance.

Default: false

runInstances_placement :: Lens' RunInstances (Maybe Placement)

The placement for the instance.

runInstances_userData :: Lens' RunInstances (Maybe Text)

The user data to make available to the instance. For more information, see Running commands on your Linux instance at launch (Linux) and Adding User Data (Windows). If you are using a command line tool, base64-encoding is performed for you, and you can load the text from a file. Otherwise, you must provide base64-encoded text. User data is limited to 16 KB.

runInstances_instanceType :: Lens' RunInstances (Maybe InstanceType)

The instance type. For more information, see Instance types in the Amazon EC2 User Guide.

Default: m1.small

runInstances_ipv6Addresses :: Lens' RunInstances (Maybe [InstanceIpv6Address])

EC2-VPC

The IPv6 addresses from the range of the subnet to associate with the primary network interface. You cannot specify this option and the option to assign a number of IPv6 addresses in the same request. You cannot specify this option if you've specified a minimum number of instances to launch.

You cannot specify this option and the network interfaces option in the same request.

runInstances_ramdiskId :: Lens' RunInstances (Maybe Text)

The ID of the RAM disk to select. Some kernels require additional drivers at launch. Check the kernel requirements for information about whether you need to specify a RAM disk. To find kernel requirements, go to the Amazon Web Services Resource Center and search for the kernel ID.

We recommend that you use PV-GRUB instead of kernels and RAM disks. For more information, see PV-GRUB in the Amazon EC2 User Guide.

runInstances_dryRun :: Lens' RunInstances (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

runInstances_creditSpecification :: Lens' RunInstances (Maybe CreditSpecificationRequest)

The credit option for CPU usage of the burstable performance instance. Valid values are standard and unlimited. To change this attribute after launch, use ModifyInstanceCreditSpecification. For more information, see Burstable performance instances in the Amazon EC2 User Guide.

Default: standard (T2 instances) or unlimited (T3/T3a instances)

For T3 instances with host tenancy, only standard is supported.

runInstances_licenseSpecifications :: Lens' RunInstances (Maybe [LicenseConfigurationRequest])

The license configurations.

runInstances_instanceMarketOptions :: Lens' RunInstances (Maybe InstanceMarketOptionsRequest)

The market (purchasing) option for the instances.

For RunInstances, persistent Spot Instance requests are only supported when InstanceInterruptionBehavior is set to either hibernate or stop.

runInstances_launchTemplate :: Lens' RunInstances (Maybe LaunchTemplateSpecification)

The launch template to use to launch the instances. Any parameters that you specify in RunInstances override the same parameters in the launch template. You can specify either the name or ID of a launch template, but not both.

runInstances_instanceInitiatedShutdownBehavior :: Lens' RunInstances (Maybe ShutdownBehavior)

Indicates whether an instance stops or terminates when you initiate shutdown from the instance (using the operating system command for system shutdown).

Default: stop

runInstances_elasticInferenceAccelerators :: Lens' RunInstances (Maybe [ElasticInferenceAccelerator])

An elastic inference accelerator to associate with the instance. Elastic inference accelerators are a resource you can attach to your Amazon EC2 instances to accelerate your Deep Learning (DL) inference workloads.

You cannot specify accelerators from different generations in the same request.

runInstances_imageId :: Lens' RunInstances (Maybe Text)

The ID of the AMI. An AMI ID is required to launch an instance and must be specified here or in a launch template.

runInstances_securityGroups :: Lens' RunInstances (Maybe [Text])

EC2-Classic, default VPC

The names of the security groups. For a nondefault VPC, you must use security group IDs instead.

If you specify a network interface, you must specify any security groups as part of the network interface.

Default: Amazon EC2 uses the default security group.

runInstances_elasticGpuSpecification :: Lens' RunInstances (Maybe [ElasticGpuSpecification])

An elastic GPU to associate with the instance. An Elastic GPU is a GPU resource that you can attach to your Windows instance to accelerate the graphics performance of your applications. For more information, see Amazon EC2 Elastic GPUs in the Amazon EC2 User Guide.

runInstances_iamInstanceProfile :: Lens' RunInstances (Maybe IamInstanceProfileSpecification)

The name or Amazon Resource Name (ARN) of an IAM instance profile.

runInstances_hibernationOptions :: Lens' RunInstances (Maybe HibernationOptionsRequest)

Indicates whether an instance is enabled for hibernation. For more information, see Hibernate your instance in the Amazon EC2 User Guide.

You can't enable hibernation and Amazon Web Services Nitro Enclaves on the same instance.

runInstances_ipv6AddressCount :: Lens' RunInstances (Maybe Int)

EC2-VPC

The number of IPv6 addresses to associate with the primary network interface. Amazon EC2 chooses the IPv6 addresses from the range of your subnet. You cannot specify this option and the option to assign specific IPv6 addresses in the same request. You can specify this option if you've specified a minimum number of instances to launch.

You cannot specify this option and the network interfaces option in the same request.

runInstances_monitoring :: Lens' RunInstances (Maybe RunInstancesMonitoringEnabled)

Specifies whether detailed monitoring is enabled for the instance.

runInstances_blockDeviceMappings :: Lens' RunInstances (Maybe [BlockDeviceMapping])

The block device mapping, which defines the EBS volumes and instance store volumes to attach to the instance at launch. For more information, see Block device mappings in the Amazon EC2 User Guide.

runInstances_subnetId :: Lens' RunInstances (Maybe Text)

EC2-VPC

The ID of the subnet to launch the instance into.

If you specify a network interface, you must specify any subnets as part of the network interface.

runInstances_enclaveOptions :: Lens' RunInstances (Maybe EnclaveOptionsRequest)

Indicates whether the instance is enabled for Amazon Web Services Nitro Enclaves. For more information, see What is Amazon Web Services Nitro Enclaves? in the Amazon Web Services Nitro Enclaves User Guide.

You can't enable Amazon Web Services Nitro Enclaves and hibernation on the same instance.

runInstances_kernelId :: Lens' RunInstances (Maybe Text)

The ID of the kernel.

We recommend that you use PV-GRUB instead of kernels and RAM disks. For more information, see PV-GRUB in the Amazon EC2 User Guide.

runInstances_cpuOptions :: Lens' RunInstances (Maybe CpuOptionsRequest)

The CPU options for the instance. For more information, see Optimizing CPU options in the Amazon EC2 User Guide.

runInstances_disableApiTermination :: Lens' RunInstances (Maybe Bool)

If you set this parameter to true, you can't terminate the instance using the Amazon EC2 console, CLI, or API; otherwise, you can. To change this attribute after launch, use ModifyInstanceAttribute. Alternatively, if you set InstanceInitiatedShutdownBehavior to terminate, you can terminate the instance by running the shutdown command from the instance.

Default: false

runInstances_networkInterfaces :: Lens' RunInstances (Maybe [InstanceNetworkInterfaceSpecification])

The network interfaces to associate with the instance. If you specify a network interface, you must specify any security groups and subnets as part of the network interface.

runInstances_keyName :: Lens' RunInstances (Maybe Text)

The name of the key pair. You can create a key pair using CreateKeyPair or ImportKeyPair.

If you do not specify a key pair, you can't connect to the instance unless you choose an AMI that is configured to allow users another way to log in.

runInstances_metadataOptions :: Lens' RunInstances (Maybe InstanceMetadataOptionsRequest)

The metadata options for the instance. For more information, see Instance metadata and user data.

runInstances_privateIpAddress :: Lens' RunInstances (Maybe Text)

EC2-VPC

The primary IPv4 address. You must specify a value from the IPv4 address range of the subnet.

Only one private IP address can be designated as primary. You can't specify this option if you've specified the option to designate a private IP address as the primary IP address in a network interface specification. You cannot specify this option if you're launching more than one instance in the request.

You cannot specify this option and the network interfaces option in the same request.

runInstances_clientToken :: Lens' RunInstances (Maybe Text)

Unique, case-sensitive identifier you provide to ensure the idempotency of the request. If you do not specify a client token, a randomly generated token is used for the request to ensure idempotency.

For more information, see Ensuring Idempotency.

Constraints: Maximum 64 ASCII characters

runInstances_maxCount :: Lens' RunInstances Int

The maximum number of instances to launch. If you specify more instances than Amazon EC2 can launch in the target Availability Zone, Amazon EC2 launches the largest possible number of instances above MinCount.

Constraints: Between 1 and the maximum number you're allowed for the specified instance type. For more information about the default limits, and how to request an increase, see How many instances can I run in Amazon EC2 in the Amazon EC2 FAQ.

runInstances_minCount :: Lens' RunInstances Int

The minimum number of instances to launch. If you specify a minimum that is more instances than Amazon EC2 can launch in the target Availability Zone, Amazon EC2 launches no instances.

Constraints: Between 1 and the maximum number you're allowed for the specified instance type. For more information about the default limits, and how to request an increase, see How many instances can I run in Amazon EC2 in the Amazon EC2 General FAQ.

Destructuring the Response

data Reservation

Describes a launch request for one or more instances, and includes owner, requester, and security group information that applies to all instances in the launch request.

See: newReservation smart constructor.

Constructors

Reservation'

Fields

groups :: Maybe [GroupIdentifier]

EC2-Classic only

The security groups.

requesterId :: Maybe Text

The ID of the requester that launched the instances on your behalf (for example, Amazon Web Services Management Console or Auto Scaling).

instances :: Maybe [Instance]

The instances.

reservationId :: Text

The ID of the reservation.

ownerId :: Text

The ID of the Amazon Web Services account that owns the reservation.

newReservation

Arguments

:: Text
$sel:reservationId:Reservation'

-> Text
$sel:ownerId:Reservation'

-> Reservation

Create a value of Reservation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groups:Reservation', reservation_groups - [EC2-Classic only] The security groups.

$sel:requesterId:Reservation', reservation_requesterId - The ID of the requester that launched the instances on your behalf (for example, Amazon Web Services Management Console or Auto Scaling).

$sel:instances:Reservation', reservation_instances - The instances.

$sel:reservationId:Reservation', reservation_reservationId - The ID of the reservation.

$sel:ownerId:Reservation', reservation_ownerId - The ID of the Amazon Web Services account that owns the reservation.

Response Lenses

reservation_groups :: Lens' Reservation (Maybe [GroupIdentifier])

EC2-Classic only

The security groups.

reservation_requesterId :: Lens' Reservation (Maybe Text)

The ID of the requester that launched the instances on your behalf (for example, Amazon Web Services Management Console or Auto Scaling).

reservation_instances :: Lens' Reservation (Maybe [Instance])

The instances.

reservation_reservationId :: Lens' Reservation Text

The ID of the reservation.

reservation_ownerId :: Lens' Reservation Text

The ID of the Amazon Web Services account that owns the reservation.