Network.AWS.EC2.SearchTransitGatewayMulticastGroups

Description

Searches one or more transit gateway multicast groups and returns the group membership information.

This operation returns paginated results.

Creating a Request

data SearchTransitGatewayMulticastGroups

See: newSearchTransitGatewayMulticastGroups smart constructor.

Constructors

SearchTransitGatewayMulticastGroups'

Fields

nextToken :: Maybe Text

The token for the next page of results.

maxResults :: Maybe Natural

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

transitGatewayMulticastDomainId :: Maybe Text

The ID of the transit gateway multicast domain.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

filters :: Maybe [Filter]

One or more filters. The possible values are:

group-ip-address - The IP address of the transit gateway multicast group.

is-group-member - The resource is a group member. Valid values are true | false.

is-group-source - The resource is a group source. Valid values are true | false.

member-type - The member type. Valid values are igmp | static.

resource-id - The ID of the resource.

resource-type - The type of resource. Valid values are vpc | vpn | direct-connect-gateway | tgw-peering.

source-type - The source type. Valid values are igmp | static.

subnet-id - The ID of the subnet.

transit-gateway-attachment-id - The id of the transit gateway attachment.

newSearchTransitGatewayMulticastGroups :: SearchTransitGatewayMulticastGroups

Create a value of SearchTransitGatewayMulticastGroups with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:SearchTransitGatewayMulticastGroups', searchTransitGatewayMulticastGroups_nextToken - The token for the next page of results.

$sel:maxResults:SearchTransitGatewayMulticastGroups', searchTransitGatewayMulticastGroups_maxResults - The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

$sel:transitGatewayMulticastDomainId:SearchTransitGatewayMulticastGroups', searchTransitGatewayMulticastGroups_transitGatewayMulticastDomainId - The ID of the transit gateway multicast domain.

$sel:dryRun:SearchTransitGatewayMulticastGroups', searchTransitGatewayMulticastGroups_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:filters:SearchTransitGatewayMulticastGroups', searchTransitGatewayMulticastGroups_filters - One or more filters. The possible values are:

group-ip-address - The IP address of the transit gateway multicast group.

is-group-member - The resource is a group member. Valid values are true | false.

is-group-source - The resource is a group source. Valid values are true | false.

member-type - The member type. Valid values are igmp | static.

resource-id - The ID of the resource.

resource-type - The type of resource. Valid values are vpc | vpn | direct-connect-gateway | tgw-peering.

source-type - The source type. Valid values are igmp | static.

subnet-id - The ID of the subnet.

transit-gateway-attachment-id - The id of the transit gateway attachment.

Request Lenses

searchTransitGatewayMulticastGroups_nextToken :: Lens' SearchTransitGatewayMulticastGroups (Maybe Text)

The token for the next page of results.

searchTransitGatewayMulticastGroups_maxResults :: Lens' SearchTransitGatewayMulticastGroups (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

searchTransitGatewayMulticastGroups_transitGatewayMulticastDomainId :: Lens' SearchTransitGatewayMulticastGroups (Maybe Text)

The ID of the transit gateway multicast domain.

searchTransitGatewayMulticastGroups_dryRun :: Lens' SearchTransitGatewayMulticastGroups (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

searchTransitGatewayMulticastGroups_filters :: Lens' SearchTransitGatewayMulticastGroups (Maybe [Filter])

One or more filters. The possible values are:

group-ip-address - The IP address of the transit gateway multicast group.

is-group-member - The resource is a group member. Valid values are true | false.

is-group-source - The resource is a group source. Valid values are true | false.

member-type - The member type. Valid values are igmp | static.

resource-id - The ID of the resource.

resource-type - The type of resource. Valid values are vpc | vpn | direct-connect-gateway | tgw-peering.

source-type - The source type. Valid values are igmp | static.

subnet-id - The ID of the subnet.

transit-gateway-attachment-id - The id of the transit gateway attachment.

Destructuring the Response

data SearchTransitGatewayMulticastGroupsResponse

See: newSearchTransitGatewayMulticastGroupsResponse smart constructor.

Constructors

SearchTransitGatewayMulticastGroupsResponse'

Fields

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

multicastGroups :: Maybe [TransitGatewayMulticastGroup]

Information about the transit gateway multicast group.

httpStatus :: Int

The response's http status code.

newSearchTransitGatewayMulticastGroupsResponse

Arguments

:: Int
$sel:httpStatus:SearchTransitGatewayMulticastGroupsResponse'

-> SearchTransitGatewayMulticastGroupsResponse

Create a value of SearchTransitGatewayMulticastGroupsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:SearchTransitGatewayMulticastGroups', searchTransitGatewayMulticastGroupsResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:multicastGroups:SearchTransitGatewayMulticastGroupsResponse', searchTransitGatewayMulticastGroupsResponse_multicastGroups - Information about the transit gateway multicast group.

$sel:httpStatus:SearchTransitGatewayMulticastGroupsResponse', searchTransitGatewayMulticastGroupsResponse_httpStatus - The response's http status code.

Response Lenses

searchTransitGatewayMulticastGroupsResponse_nextToken :: Lens' SearchTransitGatewayMulticastGroupsResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

searchTransitGatewayMulticastGroupsResponse_multicastGroups :: Lens' SearchTransitGatewayMulticastGroupsResponse (Maybe [TransitGatewayMulticastGroup])

Information about the transit gateway multicast group.

searchTransitGatewayMulticastGroupsResponse_httpStatus :: Lens' SearchTransitGatewayMulticastGroupsResponse Int

The response's http status code.