Network.AWS.EC2.SearchTransitGatewayRoutes

Description

Searches for routes in the specified transit gateway route table.

Creating a Request

data SearchTransitGatewayRoutes

See: newSearchTransitGatewayRoutes smart constructor.

Constructors

SearchTransitGatewayRoutes'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

maxResults :: Maybe Natural

The maximum number of routes to return.

transitGatewayRouteTableId :: Text

The ID of the transit gateway route table.

filters :: [Filter]

One or more filters. The possible values are:

attachment.transit-gateway-attachment-id- The id of the transit gateway attachment.

attachment.resource-id - The resource id of the transit gateway attachment.

attachment.resource-type - The attachment resource type. Valid values are vpc | vpn | direct-connect-gateway | peering | connect.

prefix-list-id - The ID of the prefix list.

route-search.exact-match - The exact match of the specified filter.

route-search.longest-prefix-match - The longest prefix that matches the route.

route-search.subnet-of-match - The routes with a subnet that match the specified CIDR filter.

route-search.supernet-of-match - The routes with a CIDR that encompass the CIDR filter. For example, if you have 10.0.1.0/29 and 10.0.1.0/31 routes in your route table and you specify supernet-of-match as 10.0.1.0/30, then the result returns 10.0.1.0/29.

state - The state of the route (active | blackhole).

type - The type of route (propagated | static).

newSearchTransitGatewayRoutes

Arguments

:: Text
$sel:transitGatewayRouteTableId:SearchTransitGatewayRoutes'

-> SearchTransitGatewayRoutes

Create a value of SearchTransitGatewayRoutes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:SearchTransitGatewayRoutes', searchTransitGatewayRoutes_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:maxResults:SearchTransitGatewayRoutes', searchTransitGatewayRoutes_maxResults - The maximum number of routes to return.

$sel:transitGatewayRouteTableId:SearchTransitGatewayRoutes', searchTransitGatewayRoutes_transitGatewayRouteTableId - The ID of the transit gateway route table.

$sel:filters:SearchTransitGatewayRoutes', searchTransitGatewayRoutes_filters - One or more filters. The possible values are:

attachment.transit-gateway-attachment-id- The id of the transit gateway attachment.

attachment.resource-id - The resource id of the transit gateway attachment.

attachment.resource-type - The attachment resource type. Valid values are vpc | vpn | direct-connect-gateway | peering | connect.

prefix-list-id - The ID of the prefix list.

route-search.exact-match - The exact match of the specified filter.

route-search.longest-prefix-match - The longest prefix that matches the route.

route-search.subnet-of-match - The routes with a subnet that match the specified CIDR filter.

route-search.supernet-of-match - The routes with a CIDR that encompass the CIDR filter. For example, if you have 10.0.1.0/29 and 10.0.1.0/31 routes in your route table and you specify supernet-of-match as 10.0.1.0/30, then the result returns 10.0.1.0/29.

state - The state of the route (active | blackhole).

type - The type of route (propagated | static).

Request Lenses

searchTransitGatewayRoutes_dryRun :: Lens' SearchTransitGatewayRoutes (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

searchTransitGatewayRoutes_maxResults :: Lens' SearchTransitGatewayRoutes (Maybe Natural)

The maximum number of routes to return.

searchTransitGatewayRoutes_transitGatewayRouteTableId :: Lens' SearchTransitGatewayRoutes Text

The ID of the transit gateway route table.

searchTransitGatewayRoutes_filters :: Lens' SearchTransitGatewayRoutes [Filter]

One or more filters. The possible values are:

attachment.transit-gateway-attachment-id- The id of the transit gateway attachment.

attachment.resource-id - The resource id of the transit gateway attachment.

attachment.resource-type - The attachment resource type. Valid values are vpc | vpn | direct-connect-gateway | peering | connect.

prefix-list-id - The ID of the prefix list.

route-search.exact-match - The exact match of the specified filter.

route-search.longest-prefix-match - The longest prefix that matches the route.

route-search.subnet-of-match - The routes with a subnet that match the specified CIDR filter.

route-search.supernet-of-match - The routes with a CIDR that encompass the CIDR filter. For example, if you have 10.0.1.0/29 and 10.0.1.0/31 routes in your route table and you specify supernet-of-match as 10.0.1.0/30, then the result returns 10.0.1.0/29.

state - The state of the route (active | blackhole).

type - The type of route (propagated | static).

Destructuring the Response

data SearchTransitGatewayRoutesResponse

See: newSearchTransitGatewayRoutesResponse smart constructor.

Constructors

SearchTransitGatewayRoutesResponse'

Fields

additionalRoutesAvailable :: Maybe Bool

Indicates whether there are additional routes available.

routes :: Maybe [TransitGatewayRoute]

Information about the routes.

httpStatus :: Int

The response's http status code.

newSearchTransitGatewayRoutesResponse

Arguments

:: Int
$sel:httpStatus:SearchTransitGatewayRoutesResponse'

-> SearchTransitGatewayRoutesResponse

Create a value of SearchTransitGatewayRoutesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:additionalRoutesAvailable:SearchTransitGatewayRoutesResponse', searchTransitGatewayRoutesResponse_additionalRoutesAvailable - Indicates whether there are additional routes available.

$sel:routes:SearchTransitGatewayRoutesResponse', searchTransitGatewayRoutesResponse_routes - Information about the routes.

$sel:httpStatus:SearchTransitGatewayRoutesResponse', searchTransitGatewayRoutesResponse_httpStatus - The response's http status code.

Response Lenses

searchTransitGatewayRoutesResponse_additionalRoutesAvailable :: Lens' SearchTransitGatewayRoutesResponse (Maybe Bool)

Indicates whether there are additional routes available.

searchTransitGatewayRoutesResponse_routes :: Lens' SearchTransitGatewayRoutesResponse (Maybe [TransitGatewayRoute])

Information about the routes.

searchTransitGatewayRoutesResponse_httpStatus :: Lens' SearchTransitGatewayRoutesResponse Int

The response's http status code.