Network.AWS.EC2.SendDiagnosticInterrupt

Description

Sends a diagnostic interrupt to the specified Amazon EC2 instance to trigger a kernel panic (on Linux instances), or a blue screen//stop error/ (on Windows instances). For instances based on Intel and AMD processors, the interrupt is received as a non-maskable interrupt (NMI).

In general, the operating system crashes and reboots when a kernel panic or stop error is triggered. The operating system can also be configured to perform diagnostic tasks, such as generating a memory dump file, loading a secondary kernel, or obtaining a call trace.

Before sending a diagnostic interrupt to your instance, ensure that its operating system is configured to perform the required diagnostic tasks.

For more information about configuring your operating system to generate a crash dump when a kernel panic or stop error occurs, see Send a diagnostic interrupt (Linux instances) or Send a Diagnostic Interrupt (Windows instances).

Creating a Request

data SendDiagnosticInterrupt

See: newSendDiagnosticInterrupt smart constructor.

Constructors

SendDiagnosticInterrupt'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

instanceId :: Text

The ID of the instance.

newSendDiagnosticInterrupt

Arguments

:: Text
$sel:instanceId:SendDiagnosticInterrupt'

-> SendDiagnosticInterrupt

Create a value of SendDiagnosticInterrupt with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:SendDiagnosticInterrupt', sendDiagnosticInterrupt_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:instanceId:SendDiagnosticInterrupt', sendDiagnosticInterrupt_instanceId - The ID of the instance.

Request Lenses

sendDiagnosticInterrupt_dryRun :: Lens' SendDiagnosticInterrupt (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

sendDiagnosticInterrupt_instanceId :: Lens' SendDiagnosticInterrupt Text

The ID of the instance.

Destructuring the Response

data SendDiagnosticInterruptResponse

See: newSendDiagnosticInterruptResponse smart constructor.

Constructors

SendDiagnosticInterruptResponse'

newSendDiagnosticInterruptResponse :: SendDiagnosticInterruptResponse

Create a value of SendDiagnosticInterruptResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.