Network.AWS.EC2.StopInstances

Description

Stops an Amazon EBS-backed instance.

You can use the Stop action to hibernate an instance if the instance is enabled for hibernation and it meets the hibernation prerequisites. For more information, see Hibernate your instance in the Amazon EC2 User Guide.

We don't charge usage for a stopped instance, or data transfer fees; however, your root partition Amazon EBS volume remains and continues to persist your data, and you are charged for Amazon EBS volume usage. Every time you start your instance, Amazon EC2 charges a one-minute minimum for instance usage, and thereafter charges per second for instance usage.

You can't stop or hibernate instance store-backed instances. You can't use the Stop action to hibernate Spot Instances, but you can specify that Amazon EC2 should hibernate Spot Instances when they are interrupted. For more information, see Hibernating interrupted Spot Instances in the Amazon EC2 User Guide.

When you stop or hibernate an instance, we shut it down. You can restart your instance at any time. Before stopping or hibernating an instance, make sure it is in a state from which it can be restarted. Stopping an instance does not preserve data stored in RAM, but hibernating an instance does preserve data stored in RAM. If an instance cannot hibernate successfully, a normal shutdown occurs.

Stopping and hibernating an instance is different to rebooting or terminating it. For example, when you stop or hibernate an instance, the root device and any other devices attached to the instance persist. When you terminate an instance, the root device and any other devices attached during the instance launch are automatically deleted. For more information about the differences between rebooting, stopping, hibernating, and terminating instances, see Instance lifecycle in the Amazon EC2 User Guide.

When you stop an instance, we attempt to shut it down forcibly after a short while. If your instance appears stuck in the stopping state after a period of time, there may be an issue with the underlying host computer. For more information, see Troubleshooting stopping your instance in the Amazon EC2 User Guide.

Creating a Request

data StopInstances

See: newStopInstances smart constructor.

Constructors

StopInstances'

Fields

hibernate :: Maybe Bool

Hibernates the instance if the instance was enabled for hibernation at launch. If the instance cannot hibernate successfully, a normal shutdown occurs. For more information, see Hibernate your instance in the Amazon EC2 User Guide.

Default: false

force :: Maybe Bool

Forces the instances to stop. The instances do not have an opportunity to flush file system caches or file system metadata. If you use this option, you must perform file system check and repair procedures. This option is not recommended for Windows instances.

Default: false

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

instanceIds :: [Text]

The IDs of the instances.

newStopInstances :: StopInstances

Create a value of StopInstances with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hibernate:StopInstances', stopInstances_hibernate - Hibernates the instance if the instance was enabled for hibernation at launch. If the instance cannot hibernate successfully, a normal shutdown occurs. For more information, see Hibernate your instance in the Amazon EC2 User Guide.

Default: false

$sel:force:StopInstances', stopInstances_force - Forces the instances to stop. The instances do not have an opportunity to flush file system caches or file system metadata. If you use this option, you must perform file system check and repair procedures. This option is not recommended for Windows instances.

Default: false

$sel:dryRun:StopInstances', stopInstances_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:instanceIds:StopInstances', stopInstances_instanceIds - The IDs of the instances.

Request Lenses

stopInstances_hibernate :: Lens' StopInstances (Maybe Bool)

Hibernates the instance if the instance was enabled for hibernation at launch. If the instance cannot hibernate successfully, a normal shutdown occurs. For more information, see Hibernate your instance in the Amazon EC2 User Guide.

Default: false

stopInstances_force :: Lens' StopInstances (Maybe Bool)

Forces the instances to stop. The instances do not have an opportunity to flush file system caches or file system metadata. If you use this option, you must perform file system check and repair procedures. This option is not recommended for Windows instances.

Default: false

stopInstances_dryRun :: Lens' StopInstances (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

stopInstances_instanceIds :: Lens' StopInstances [Text]

The IDs of the instances.

Destructuring the Response

data StopInstancesResponse

See: newStopInstancesResponse smart constructor.

Constructors

StopInstancesResponse'

Fields

stoppingInstances :: Maybe [InstanceStateChange]

Information about the stopped instances.

httpStatus :: Int

The response's http status code.

newStopInstancesResponse

Arguments

:: Int
$sel:httpStatus:StopInstancesResponse'

-> StopInstancesResponse

Create a value of StopInstancesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stoppingInstances:StopInstancesResponse', stopInstancesResponse_stoppingInstances - Information about the stopped instances.

$sel:httpStatus:StopInstancesResponse', stopInstancesResponse_httpStatus - The response's http status code.

Response Lenses

stopInstancesResponse_stoppingInstances :: Lens' StopInstancesResponse (Maybe [InstanceStateChange])

Information about the stopped instances.

stopInstancesResponse_httpStatus :: Lens' StopInstancesResponse Int

The response's http status code.