Network.AWS.EC2.Types.Address

Description

Documentation

data Address

Describes an Elastic IP address, or a carrier IP address.

See: newAddress smart constructor.

Constructors

Address'

Fields

instanceId :: Maybe Text

The ID of the instance that the address is associated with (if any).

customerOwnedIpv4Pool :: Maybe Text

The ID of the customer-owned address pool.

domain :: Maybe DomainType

Indicates whether this Elastic IP address is for use with instances in EC2-Classic (standard) or instances in a VPC (vpc).

carrierIp :: Maybe Text

The carrier IP address associated. This option is only available for network interfaces which reside in a subnet in a Wavelength Zone (for example an EC2 instance).

customerOwnedIp :: Maybe Text

The customer-owned IP address.

networkInterfaceOwnerId :: Maybe Text

The ID of the Amazon Web Services account that owns the network interface.

associationId :: Maybe Text

The ID representing the association of the address with an instance in a VPC.

tags :: Maybe [Tag]

Any tags assigned to the Elastic IP address.

publicIpv4Pool :: Maybe Text

The ID of an address pool.

networkInterfaceId :: Maybe Text

The ID of the network interface.

publicIp :: Maybe Text

The Elastic IP address.

allocationId :: Maybe Text

The ID representing the allocation of the address for use with EC2-VPC.

networkBorderGroup :: Maybe Text

The name of the unique set of Availability Zones, Local Zones, or Wavelength Zones from which AWS advertises IP addresses.

privateIpAddress :: Maybe Text

The private IP address associated with the Elastic IP address.

newAddress :: Address

Create a value of Address with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceId:Address', address_instanceId - The ID of the instance that the address is associated with (if any).

$sel:customerOwnedIpv4Pool:Address', address_customerOwnedIpv4Pool - The ID of the customer-owned address pool.

$sel:domain:Address', address_domain - Indicates whether this Elastic IP address is for use with instances in EC2-Classic (standard) or instances in a VPC (vpc).

$sel:carrierIp:Address', address_carrierIp - The carrier IP address associated. This option is only available for network interfaces which reside in a subnet in a Wavelength Zone (for example an EC2 instance).

$sel:customerOwnedIp:Address', address_customerOwnedIp - The customer-owned IP address.

$sel:networkInterfaceOwnerId:Address', address_networkInterfaceOwnerId - The ID of the Amazon Web Services account that owns the network interface.

$sel:associationId:Address', address_associationId - The ID representing the association of the address with an instance in a VPC.

$sel:tags:Address', address_tags - Any tags assigned to the Elastic IP address.

$sel:publicIpv4Pool:Address', address_publicIpv4Pool - The ID of an address pool.

$sel:networkInterfaceId:Address', address_networkInterfaceId - The ID of the network interface.

$sel:publicIp:Address', address_publicIp - The Elastic IP address.

$sel:allocationId:Address', address_allocationId - The ID representing the allocation of the address for use with EC2-VPC.

$sel:networkBorderGroup:Address', address_networkBorderGroup - The name of the unique set of Availability Zones, Local Zones, or Wavelength Zones from which AWS advertises IP addresses.

$sel:privateIpAddress:Address', address_privateIpAddress - The private IP address associated with the Elastic IP address.

address_instanceId :: Lens' Address (Maybe Text)

The ID of the instance that the address is associated with (if any).

address_customerOwnedIpv4Pool :: Lens' Address (Maybe Text)

The ID of the customer-owned address pool.

address_domain :: Lens' Address (Maybe DomainType)

Indicates whether this Elastic IP address is for use with instances in EC2-Classic (standard) or instances in a VPC (vpc).

address_carrierIp :: Lens' Address (Maybe Text)

The carrier IP address associated. This option is only available for network interfaces which reside in a subnet in a Wavelength Zone (for example an EC2 instance).

address_customerOwnedIp :: Lens' Address (Maybe Text)

The customer-owned IP address.

address_networkInterfaceOwnerId :: Lens' Address (Maybe Text)

The ID of the Amazon Web Services account that owns the network interface.

address_associationId :: Lens' Address (Maybe Text)

The ID representing the association of the address with an instance in a VPC.

address_tags :: Lens' Address (Maybe [Tag])

Any tags assigned to the Elastic IP address.

address_publicIpv4Pool :: Lens' Address (Maybe Text)

The ID of an address pool.

address_networkInterfaceId :: Lens' Address (Maybe Text)

The ID of the network interface.

address_publicIp :: Lens' Address (Maybe Text)

The Elastic IP address.

address_allocationId :: Lens' Address (Maybe Text)

The ID representing the allocation of the address for use with EC2-VPC.

address_networkBorderGroup :: Lens' Address (Maybe Text)

The name of the unique set of Availability Zones, Local Zones, or Wavelength Zones from which AWS advertises IP addresses.

address_privateIpAddress :: Lens' Address (Maybe Text)

The private IP address associated with the Elastic IP address.