Network.AWS.EC2.Types.AnalysisRouteTableRoute

Description

Documentation

data AnalysisRouteTableRoute

Describes a route table route.

See: newAnalysisRouteTableRoute smart constructor.

Constructors

AnalysisRouteTableRoute'

Fields

instanceId :: Maybe Text

The ID of the instance, such as a NAT instance.

origin :: Maybe Text

Describes how the route was created. The following are possible values:

CreateRouteTable - The route was automatically created when the route table was created.

CreateRoute - The route was manually added to the route table.

EnableVgwRoutePropagation - The route was propagated by route propagation.

vpcPeeringConnectionId :: Maybe Text

The ID of a VPC peering connection.

destinationPrefixListId :: Maybe Text

The prefix of the Amazon Web Service.

destinationCidr :: Maybe Text

The destination IPv4 address, in CIDR notation.

egressOnlyInternetGatewayId :: Maybe Text

The ID of an egress-only internet gateway.

natGatewayId :: Maybe Text

The ID of a NAT gateway.

networkInterfaceId :: Maybe Text

The ID of a network interface.

gatewayId :: Maybe Text

The ID of the gateway, such as an internet gateway or virtual private gateway.

transitGatewayId :: Maybe Text

The ID of a transit gateway.

newAnalysisRouteTableRoute :: AnalysisRouteTableRoute

Create a value of AnalysisRouteTableRoute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceId:AnalysisRouteTableRoute', analysisRouteTableRoute_instanceId - The ID of the instance, such as a NAT instance.

$sel:origin:AnalysisRouteTableRoute', analysisRouteTableRoute_origin - Describes how the route was created. The following are possible values:

CreateRouteTable - The route was automatically created when the route table was created.

CreateRoute - The route was manually added to the route table.

EnableVgwRoutePropagation - The route was propagated by route propagation.

$sel:vpcPeeringConnectionId:AnalysisRouteTableRoute', analysisRouteTableRoute_vpcPeeringConnectionId - The ID of a VPC peering connection.

$sel:destinationPrefixListId:AnalysisRouteTableRoute', analysisRouteTableRoute_destinationPrefixListId - The prefix of the Amazon Web Service.

$sel:destinationCidr:AnalysisRouteTableRoute', analysisRouteTableRoute_destinationCidr - The destination IPv4 address, in CIDR notation.

$sel:egressOnlyInternetGatewayId:AnalysisRouteTableRoute', analysisRouteTableRoute_egressOnlyInternetGatewayId - The ID of an egress-only internet gateway.

$sel:natGatewayId:AnalysisRouteTableRoute', analysisRouteTableRoute_natGatewayId - The ID of a NAT gateway.

$sel:networkInterfaceId:AnalysisRouteTableRoute', analysisRouteTableRoute_networkInterfaceId - The ID of a network interface.

$sel:gatewayId:AnalysisRouteTableRoute', analysisRouteTableRoute_gatewayId - The ID of the gateway, such as an internet gateway or virtual private gateway.

$sel:transitGatewayId:AnalysisRouteTableRoute', analysisRouteTableRoute_transitGatewayId - The ID of a transit gateway.

analysisRouteTableRoute_instanceId :: Lens' AnalysisRouteTableRoute (Maybe Text)

The ID of the instance, such as a NAT instance.

analysisRouteTableRoute_origin :: Lens' AnalysisRouteTableRoute (Maybe Text)

Describes how the route was created. The following are possible values:

CreateRouteTable - The route was automatically created when the route table was created.

CreateRoute - The route was manually added to the route table.

EnableVgwRoutePropagation - The route was propagated by route propagation.

analysisRouteTableRoute_vpcPeeringConnectionId :: Lens' AnalysisRouteTableRoute (Maybe Text)

The ID of a VPC peering connection.

analysisRouteTableRoute_destinationPrefixListId :: Lens' AnalysisRouteTableRoute (Maybe Text)

The prefix of the Amazon Web Service.

analysisRouteTableRoute_destinationCidr :: Lens' AnalysisRouteTableRoute (Maybe Text)

The destination IPv4 address, in CIDR notation.

analysisRouteTableRoute_egressOnlyInternetGatewayId :: Lens' AnalysisRouteTableRoute (Maybe Text)

The ID of an egress-only internet gateway.

analysisRouteTableRoute_natGatewayId :: Lens' AnalysisRouteTableRoute (Maybe Text)

The ID of a NAT gateway.

analysisRouteTableRoute_networkInterfaceId :: Lens' AnalysisRouteTableRoute (Maybe Text)

The ID of a network interface.

analysisRouteTableRoute_gatewayId :: Lens' AnalysisRouteTableRoute (Maybe Text)

The ID of the gateway, such as an internet gateway or virtual private gateway.

analysisRouteTableRoute_transitGatewayId :: Lens' AnalysisRouteTableRoute (Maybe Text)

The ID of a transit gateway.