Network.AWS.EC2.Types.AvailabilityZone

Description

Documentation

data AvailabilityZone

Describes Availability Zones, Local Zones, and Wavelength Zones.

See: newAvailabilityZone smart constructor.

Constructors

AvailabilityZone'

Fields

parentZoneId :: Maybe Text

The ID of the zone that handles some of the Local Zone or Wavelength Zone control plane operations, such as API calls.

regionName :: Maybe Text

The name of the Region.

zoneName :: Maybe Text

The name of the Availability Zone, Local Zone, or Wavelength Zone.

zoneType :: Maybe Text

The type of zone. The valid values are availability-zone, local-zone, and wavelength-zone.

zoneId :: Maybe Text

The ID of the Availability Zone, Local Zone, or Wavelength Zone.

optInStatus :: Maybe AvailabilityZoneOptInStatus

For Availability Zones, this parameter always has the value of opt-in-not-required.

For Local Zones and Wavelength Zones, this parameter is the opt-in status. The possible values are opted-in, and not-opted-in.

groupName :: Maybe Text

For Availability Zones, this parameter has the same value as the Region name.

For Local Zones, the name of the associated group, for example us-west-2-lax-1.

For Wavelength Zones, the name of the associated group, for example us-east-1-wl1-bos-wlz-1.

state :: Maybe AvailabilityZoneState

The state of the Availability Zone, Local Zone, or Wavelength Zone.

messages :: Maybe [AvailabilityZoneMessage]

Any messages about the Availability Zone, Local Zone, or Wavelength Zone.

parentZoneName :: Maybe Text

The name of the zone that handles some of the Local Zone or Wavelength Zone control plane operations, such as API calls.

networkBorderGroup :: Maybe Text

The name of the network border group.

newAvailabilityZone :: AvailabilityZone

Create a value of AvailabilityZone with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parentZoneId:AvailabilityZone', availabilityZone_parentZoneId - The ID of the zone that handles some of the Local Zone or Wavelength Zone control plane operations, such as API calls.

$sel:regionName:AvailabilityZone', availabilityZone_regionName - The name of the Region.

$sel:zoneName:AvailabilityZone', availabilityZone_zoneName - The name of the Availability Zone, Local Zone, or Wavelength Zone.

$sel:zoneType:AvailabilityZone', availabilityZone_zoneType - The type of zone. The valid values are availability-zone, local-zone, and wavelength-zone.

$sel:zoneId:AvailabilityZone', availabilityZone_zoneId - The ID of the Availability Zone, Local Zone, or Wavelength Zone.

$sel:optInStatus:AvailabilityZone', availabilityZone_optInStatus - For Availability Zones, this parameter always has the value of opt-in-not-required.

For Local Zones and Wavelength Zones, this parameter is the opt-in status. The possible values are opted-in, and not-opted-in.

$sel:groupName:AvailabilityZone', availabilityZone_groupName - For Availability Zones, this parameter has the same value as the Region name.

For Local Zones, the name of the associated group, for example us-west-2-lax-1.

For Wavelength Zones, the name of the associated group, for example us-east-1-wl1-bos-wlz-1.

$sel:state:AvailabilityZone', availabilityZone_state - The state of the Availability Zone, Local Zone, or Wavelength Zone.

$sel:messages:AvailabilityZone', availabilityZone_messages - Any messages about the Availability Zone, Local Zone, or Wavelength Zone.

$sel:parentZoneName:AvailabilityZone', availabilityZone_parentZoneName - The name of the zone that handles some of the Local Zone or Wavelength Zone control plane operations, such as API calls.

$sel:networkBorderGroup:AvailabilityZone', availabilityZone_networkBorderGroup - The name of the network border group.

availabilityZone_parentZoneId :: Lens' AvailabilityZone (Maybe Text)

The ID of the zone that handles some of the Local Zone or Wavelength Zone control plane operations, such as API calls.

availabilityZone_regionName :: Lens' AvailabilityZone (Maybe Text)

The name of the Region.

availabilityZone_zoneName :: Lens' AvailabilityZone (Maybe Text)

The name of the Availability Zone, Local Zone, or Wavelength Zone.

availabilityZone_zoneType :: Lens' AvailabilityZone (Maybe Text)

The type of zone. The valid values are availability-zone, local-zone, and wavelength-zone.

availabilityZone_zoneId :: Lens' AvailabilityZone (Maybe Text)

The ID of the Availability Zone, Local Zone, or Wavelength Zone.

availabilityZone_optInStatus :: Lens' AvailabilityZone (Maybe AvailabilityZoneOptInStatus)

For Availability Zones, this parameter always has the value of opt-in-not-required.

For Local Zones and Wavelength Zones, this parameter is the opt-in status. The possible values are opted-in, and not-opted-in.

availabilityZone_groupName :: Lens' AvailabilityZone (Maybe Text)

For Availability Zones, this parameter has the same value as the Region name.

For Local Zones, the name of the associated group, for example us-west-2-lax-1.

For Wavelength Zones, the name of the associated group, for example us-east-1-wl1-bos-wlz-1.

availabilityZone_state :: Lens' AvailabilityZone (Maybe AvailabilityZoneState)

The state of the Availability Zone, Local Zone, or Wavelength Zone.

availabilityZone_messages :: Lens' AvailabilityZone (Maybe [AvailabilityZoneMessage])

Any messages about the Availability Zone, Local Zone, or Wavelength Zone.

availabilityZone_parentZoneName :: Lens' AvailabilityZone (Maybe Text)

The name of the zone that handles some of the Local Zone or Wavelength Zone control plane operations, such as API calls.

availabilityZone_networkBorderGroup :: Lens' AvailabilityZone (Maybe Text)

The name of the network border group.