Network.AWS.EC2.Types.BlockDeviceMapping

Description

Documentation

data BlockDeviceMapping

Describes a block device mapping, which defines the EBS volumes and instance store volumes to attach to an instance at launch.

See: newBlockDeviceMapping smart constructor.

Constructors

BlockDeviceMapping'

Fields

ebs :: Maybe EbsBlockDevice

Parameters used to automatically set up EBS volumes when the instance is launched.

noDevice :: Maybe Text

To omit the device from the block device mapping, specify an empty string. When this property is specified, the device is removed from the block device mapping regardless of the assigned value.

virtualName :: Maybe Text

The virtual device name (ephemeralN). Instance store volumes are numbered starting from 0. An instance type with 2 available instance store volumes can specify mappings for ephemeral0 and ephemeral1. The number of available instance store volumes depends on the instance type. After you connect to the instance, you must mount the volume.

NVMe instance store volumes are automatically enumerated and assigned a device name. Including them in your block device mapping has no effect.

Constraints: For M3 instances, you must specify instance store volumes in the block device mapping for the instance. When you launch an M3 instance, we ignore any instance store volumes specified in the block device mapping for the AMI.

deviceName :: Text

The device name (for example, /dev/sdh or xvdh).

newBlockDeviceMapping

Arguments

:: Text
$sel:deviceName:BlockDeviceMapping'

-> BlockDeviceMapping

Create a value of BlockDeviceMapping with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ebs:BlockDeviceMapping', blockDeviceMapping_ebs - Parameters used to automatically set up EBS volumes when the instance is launched.

$sel:noDevice:BlockDeviceMapping', blockDeviceMapping_noDevice - To omit the device from the block device mapping, specify an empty string. When this property is specified, the device is removed from the block device mapping regardless of the assigned value.

$sel:virtualName:BlockDeviceMapping', blockDeviceMapping_virtualName - The virtual device name (ephemeralN). Instance store volumes are numbered starting from 0. An instance type with 2 available instance store volumes can specify mappings for ephemeral0 and ephemeral1. The number of available instance store volumes depends on the instance type. After you connect to the instance, you must mount the volume.

NVMe instance store volumes are automatically enumerated and assigned a device name. Including them in your block device mapping has no effect.

Constraints: For M3 instances, you must specify instance store volumes in the block device mapping for the instance. When you launch an M3 instance, we ignore any instance store volumes specified in the block device mapping for the AMI.

$sel:deviceName:BlockDeviceMapping', blockDeviceMapping_deviceName - The device name (for example, /dev/sdh or xvdh).

blockDeviceMapping_ebs :: Lens' BlockDeviceMapping (Maybe EbsBlockDevice)

Parameters used to automatically set up EBS volumes when the instance is launched.

blockDeviceMapping_noDevice :: Lens' BlockDeviceMapping (Maybe Text)

To omit the device from the block device mapping, specify an empty string. When this property is specified, the device is removed from the block device mapping regardless of the assigned value.

blockDeviceMapping_virtualName :: Lens' BlockDeviceMapping (Maybe Text)

The virtual device name (ephemeralN). Instance store volumes are numbered starting from 0. An instance type with 2 available instance store volumes can specify mappings for ephemeral0 and ephemeral1. The number of available instance store volumes depends on the instance type. After you connect to the instance, you must mount the volume.

NVMe instance store volumes are automatically enumerated and assigned a device name. Including them in your block device mapping has no effect.

Constraints: For M3 instances, you must specify instance store volumes in the block device mapping for the instance. When you launch an M3 instance, we ignore any instance store volumes specified in the block device mapping for the AMI.

blockDeviceMapping_deviceName :: Lens' BlockDeviceMapping Text

The device name (for example, /dev/sdh or xvdh).