Network.AWS.EC2.Types.BundleTask

Description

Documentation

data BundleTask

Describes a bundle task.

See: newBundleTask smart constructor.

Constructors

BundleTask'

Fields

bundleTaskError :: Maybe BundleTaskError

If the task fails, a description of the error.

bundleId :: Text

The ID of the bundle task.

instanceId :: Text

The ID of the instance associated with this bundle task.

progress :: Text

The level of task completion, as a percent (for example, 20%).

startTime :: ISO8601

The time this task started.

state :: BundleTaskState

The state of the task.

storage :: Storage

The Amazon S3 storage locations.

updateTime :: ISO8601

The time of the most recent update for the task.

newBundleTask

Arguments

:: Text
$sel:bundleId:BundleTask'

-> Text
$sel:instanceId:BundleTask'

-> Text
$sel:progress:BundleTask'

-> UTCTime
$sel:startTime:BundleTask'

-> BundleTaskState
$sel:state:BundleTask'

-> Storage
$sel:storage:BundleTask'

-> UTCTime
$sel:updateTime:BundleTask'

-> BundleTask

Create a value of BundleTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bundleTaskError:BundleTask', bundleTask_bundleTaskError - If the task fails, a description of the error.

$sel:bundleId:BundleTask', bundleTask_bundleId - The ID of the bundle task.

$sel:instanceId:BundleTask', bundleTask_instanceId - The ID of the instance associated with this bundle task.

$sel:progress:BundleTask', bundleTask_progress - The level of task completion, as a percent (for example, 20%).

$sel:startTime:BundleTask', bundleTask_startTime - The time this task started.

$sel:state:BundleTask', bundleTask_state - The state of the task.

$sel:storage:BundleTask', bundleTask_storage - The Amazon S3 storage locations.

$sel:updateTime:BundleTask', bundleTask_updateTime - The time of the most recent update for the task.

bundleTask_bundleTaskError :: Lens' BundleTask (Maybe BundleTaskError)

If the task fails, a description of the error.

bundleTask_bundleId :: Lens' BundleTask Text

The ID of the bundle task.

bundleTask_instanceId :: Lens' BundleTask Text

The ID of the instance associated with this bundle task.

bundleTask_progress :: Lens' BundleTask Text

The level of task completion, as a percent (for example, 20%).

bundleTask_startTime :: Lens' BundleTask UTCTime

The time this task started.

bundleTask_state :: Lens' BundleTask BundleTaskState

The state of the task.

bundleTask_storage :: Lens' BundleTask Storage

The Amazon S3 storage locations.

bundleTask_updateTime :: Lens' BundleTask UTCTime

The time of the most recent update for the task.