Network.AWS.EC2.Types.CancelSpotFleetRequestsSuccessItem

Description

Documentation

data CancelSpotFleetRequestsSuccessItem

Describes a Spot Fleet request that was successfully canceled.

See: newCancelSpotFleetRequestsSuccessItem smart constructor.

Constructors

CancelSpotFleetRequestsSuccessItem'

Fields

currentSpotFleetRequestState :: Maybe BatchState

The current state of the Spot Fleet request.

spotFleetRequestId :: Maybe Text

The ID of the Spot Fleet request.

previousSpotFleetRequestState :: Maybe BatchState

The previous state of the Spot Fleet request.

newCancelSpotFleetRequestsSuccessItem :: CancelSpotFleetRequestsSuccessItem

Create a value of CancelSpotFleetRequestsSuccessItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:currentSpotFleetRequestState:CancelSpotFleetRequestsSuccessItem', cancelSpotFleetRequestsSuccessItem_currentSpotFleetRequestState - The current state of the Spot Fleet request.

$sel:spotFleetRequestId:CancelSpotFleetRequestsSuccessItem', cancelSpotFleetRequestsSuccessItem_spotFleetRequestId - The ID of the Spot Fleet request.

$sel:previousSpotFleetRequestState:CancelSpotFleetRequestsSuccessItem', cancelSpotFleetRequestsSuccessItem_previousSpotFleetRequestState - The previous state of the Spot Fleet request.

cancelSpotFleetRequestsSuccessItem_currentSpotFleetRequestState :: Lens' CancelSpotFleetRequestsSuccessItem (Maybe BatchState)

The current state of the Spot Fleet request.

cancelSpotFleetRequestsSuccessItem_spotFleetRequestId :: Lens' CancelSpotFleetRequestsSuccessItem (Maybe Text)

The ID of the Spot Fleet request.

cancelSpotFleetRequestsSuccessItem_previousSpotFleetRequestState :: Lens' CancelSpotFleetRequestsSuccessItem (Maybe BatchState)

The previous state of the Spot Fleet request.