Network.AWS.EC2.Types.CapacityReservation

Description

Documentation

data CapacityReservation

Describes a Capacity Reservation.

See: newCapacityReservation smart constructor.

Constructors

CapacityReservation'

Fields

ownerId :: Maybe Text

The ID of the Amazon Web Services account that owns the Capacity Reservation.

startDate :: Maybe ISO8601

The date and time at which the Capacity Reservation was started.

ebsOptimized :: Maybe Bool

Indicates whether the Capacity Reservation supports EBS-optimized instances. This optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal I/O performance. This optimization isn't available with all instance types. Additional usage charges apply when using an EBS- optimized instance.

instanceType :: Maybe Text

The type of instance for which the Capacity Reservation reserves capacity.

availableInstanceCount :: Maybe Int

The remaining capacity. Indicates the number of instances that can be launched in the Capacity Reservation.

outpostArn :: Maybe Text

The Amazon Resource Name (ARN) of the Outpost on which the Capacity Reservation was created.

endDateType :: Maybe EndDateType

Indicates the way in which the Capacity Reservation ends. A Capacity Reservation can have one of the following end types:

unlimited - The Capacity Reservation remains active until you explicitly cancel it.

limited - The Capacity Reservation expires automatically at a specified date and time.

createDate :: Maybe ISO8601

The date and time at which the Capacity Reservation was created.

availabilityZoneId :: Maybe Text

The Availability Zone ID of the Capacity Reservation.

tenancy :: Maybe CapacityReservationTenancy

Indicates the tenancy of the Capacity Reservation. A Capacity Reservation can have one of the following tenancy settings:

default - The Capacity Reservation is created on hardware that is shared with other Amazon Web Services accounts.

dedicated - The Capacity Reservation is created on single-tenant hardware that is dedicated to a single Amazon Web Services account.

availabilityZone :: Maybe Text

The Availability Zone in which the capacity is reserved.

state :: Maybe CapacityReservationState

The current state of the Capacity Reservation. A Capacity Reservation can be in one of the following states:

active - The Capacity Reservation is active and the capacity is available for your use.

expired - The Capacity Reservation expired automatically at the date and time specified in your request. The reserved capacity is no longer available for your use.

cancelled - The Capacity Reservation was cancelled. The reserved capacity is no longer available for your use.

pending - The Capacity Reservation request was successful but the capacity provisioning is still pending.

failed - The Capacity Reservation request has failed. A request might fail due to invalid request parameters, capacity constraints, or instance limit constraints. Failed requests are retained for 60 minutes.

tags :: Maybe [Tag]

Any tags assigned to the Capacity Reservation.

capacityReservationArn :: Maybe Text

The Amazon Resource Name (ARN) of the Capacity Reservation.

instanceMatchCriteria :: Maybe InstanceMatchCriteria

Indicates the type of instance launches that the Capacity Reservation accepts. The options include:

open - The Capacity Reservation accepts all instances that have matching attributes (instance type, platform, and Availability Zone). Instances that have matching attributes launch into the Capacity Reservation automatically without specifying any additional parameters.

targeted - The Capacity Reservation only accepts instances that have matching attributes (instance type, platform, and Availability Zone), and explicitly target the Capacity Reservation. This ensures that only permitted instances can use the reserved capacity.

capacityReservationId :: Maybe Text

The ID of the Capacity Reservation.

instancePlatform :: Maybe CapacityReservationInstancePlatform

The type of operating system for which the Capacity Reservation reserves capacity.

ephemeralStorage :: Maybe Bool

Indicates whether the Capacity Reservation supports instances with temporary, block-level storage.

endDate :: Maybe ISO8601

The date and time at which the Capacity Reservation expires. When a Capacity Reservation expires, the reserved capacity is released and you can no longer launch instances into it. The Capacity Reservation's state changes to expired when it reaches its end date and time.

totalInstanceCount :: Maybe Int

The total number of instances for which the Capacity Reservation reserves capacity.

newCapacityReservation :: CapacityReservation

Create a value of CapacityReservation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ownerId:CapacityReservation', capacityReservation_ownerId - The ID of the Amazon Web Services account that owns the Capacity Reservation.

$sel:startDate:CapacityReservation', capacityReservation_startDate - The date and time at which the Capacity Reservation was started.

$sel:ebsOptimized:CapacityReservation', capacityReservation_ebsOptimized - Indicates whether the Capacity Reservation supports EBS-optimized instances. This optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal I/O performance. This optimization isn't available with all instance types. Additional usage charges apply when using an EBS- optimized instance.

$sel:instanceType:CapacityReservation', capacityReservation_instanceType - The type of instance for which the Capacity Reservation reserves capacity.

$sel:availableInstanceCount:CapacityReservation', capacityReservation_availableInstanceCount - The remaining capacity. Indicates the number of instances that can be launched in the Capacity Reservation.

$sel:outpostArn:CapacityReservation', capacityReservation_outpostArn - The Amazon Resource Name (ARN) of the Outpost on which the Capacity Reservation was created.

$sel:endDateType:CapacityReservation', capacityReservation_endDateType - Indicates the way in which the Capacity Reservation ends. A Capacity Reservation can have one of the following end types:

unlimited - The Capacity Reservation remains active until you explicitly cancel it.

limited - The Capacity Reservation expires automatically at a specified date and time.

$sel:createDate:CapacityReservation', capacityReservation_createDate - The date and time at which the Capacity Reservation was created.

$sel:availabilityZoneId:CapacityReservation', capacityReservation_availabilityZoneId - The Availability Zone ID of the Capacity Reservation.

$sel:tenancy:CapacityReservation', capacityReservation_tenancy - Indicates the tenancy of the Capacity Reservation. A Capacity Reservation can have one of the following tenancy settings:

default - The Capacity Reservation is created on hardware that is shared with other Amazon Web Services accounts.

dedicated - The Capacity Reservation is created on single-tenant hardware that is dedicated to a single Amazon Web Services account.

$sel:availabilityZone:CapacityReservation', capacityReservation_availabilityZone - The Availability Zone in which the capacity is reserved.

$sel:state:CapacityReservation', capacityReservation_state - The current state of the Capacity Reservation. A Capacity Reservation can be in one of the following states:

active - The Capacity Reservation is active and the capacity is available for your use.

expired - The Capacity Reservation expired automatically at the date and time specified in your request. The reserved capacity is no longer available for your use.

cancelled - The Capacity Reservation was cancelled. The reserved capacity is no longer available for your use.

pending - The Capacity Reservation request was successful but the capacity provisioning is still pending.

failed - The Capacity Reservation request has failed. A request might fail due to invalid request parameters, capacity constraints, or instance limit constraints. Failed requests are retained for 60 minutes.

$sel:tags:CapacityReservation', capacityReservation_tags - Any tags assigned to the Capacity Reservation.

$sel:capacityReservationArn:CapacityReservation', capacityReservation_capacityReservationArn - The Amazon Resource Name (ARN) of the Capacity Reservation.

$sel:instanceMatchCriteria:CapacityReservation', capacityReservation_instanceMatchCriteria - Indicates the type of instance launches that the Capacity Reservation accepts. The options include:

open - The Capacity Reservation accepts all instances that have matching attributes (instance type, platform, and Availability Zone). Instances that have matching attributes launch into the Capacity Reservation automatically without specifying any additional parameters.

targeted - The Capacity Reservation only accepts instances that have matching attributes (instance type, platform, and Availability Zone), and explicitly target the Capacity Reservation. This ensures that only permitted instances can use the reserved capacity.

$sel:capacityReservationId:CapacityReservation', capacityReservation_capacityReservationId - The ID of the Capacity Reservation.

$sel:instancePlatform:CapacityReservation', capacityReservation_instancePlatform - The type of operating system for which the Capacity Reservation reserves capacity.

$sel:ephemeralStorage:CapacityReservation', capacityReservation_ephemeralStorage - Indicates whether the Capacity Reservation supports instances with temporary, block-level storage.

$sel:endDate:CapacityReservation', capacityReservation_endDate - The date and time at which the Capacity Reservation expires. When a Capacity Reservation expires, the reserved capacity is released and you can no longer launch instances into it. The Capacity Reservation's state changes to expired when it reaches its end date and time.

$sel:totalInstanceCount:CapacityReservation', capacityReservation_totalInstanceCount - The total number of instances for which the Capacity Reservation reserves capacity.

capacityReservation_ownerId :: Lens' CapacityReservation (Maybe Text)

The ID of the Amazon Web Services account that owns the Capacity Reservation.

capacityReservation_startDate :: Lens' CapacityReservation (Maybe UTCTime)

The date and time at which the Capacity Reservation was started.

capacityReservation_ebsOptimized :: Lens' CapacityReservation (Maybe Bool)

Indicates whether the Capacity Reservation supports EBS-optimized instances. This optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal I/O performance. This optimization isn't available with all instance types. Additional usage charges apply when using an EBS- optimized instance.

capacityReservation_instanceType :: Lens' CapacityReservation (Maybe Text)

The type of instance for which the Capacity Reservation reserves capacity.

capacityReservation_availableInstanceCount :: Lens' CapacityReservation (Maybe Int)

The remaining capacity. Indicates the number of instances that can be launched in the Capacity Reservation.

capacityReservation_outpostArn :: Lens' CapacityReservation (Maybe Text)

The Amazon Resource Name (ARN) of the Outpost on which the Capacity Reservation was created.

capacityReservation_endDateType :: Lens' CapacityReservation (Maybe EndDateType)

Indicates the way in which the Capacity Reservation ends. A Capacity Reservation can have one of the following end types:

unlimited - The Capacity Reservation remains active until you explicitly cancel it.

limited - The Capacity Reservation expires automatically at a specified date and time.

capacityReservation_createDate :: Lens' CapacityReservation (Maybe UTCTime)

The date and time at which the Capacity Reservation was created.

capacityReservation_availabilityZoneId :: Lens' CapacityReservation (Maybe Text)

The Availability Zone ID of the Capacity Reservation.

capacityReservation_tenancy :: Lens' CapacityReservation (Maybe CapacityReservationTenancy)

Indicates the tenancy of the Capacity Reservation. A Capacity Reservation can have one of the following tenancy settings:

default - The Capacity Reservation is created on hardware that is shared with other Amazon Web Services accounts.

dedicated - The Capacity Reservation is created on single-tenant hardware that is dedicated to a single Amazon Web Services account.

capacityReservation_availabilityZone :: Lens' CapacityReservation (Maybe Text)

The Availability Zone in which the capacity is reserved.

capacityReservation_state :: Lens' CapacityReservation (Maybe CapacityReservationState)

The current state of the Capacity Reservation. A Capacity Reservation can be in one of the following states:

active - The Capacity Reservation is active and the capacity is available for your use.

expired - The Capacity Reservation expired automatically at the date and time specified in your request. The reserved capacity is no longer available for your use.

cancelled - The Capacity Reservation was cancelled. The reserved capacity is no longer available for your use.

pending - The Capacity Reservation request was successful but the capacity provisioning is still pending.

failed - The Capacity Reservation request has failed. A request might fail due to invalid request parameters, capacity constraints, or instance limit constraints. Failed requests are retained for 60 minutes.

capacityReservation_tags :: Lens' CapacityReservation (Maybe [Tag])

Any tags assigned to the Capacity Reservation.

capacityReservation_capacityReservationArn :: Lens' CapacityReservation (Maybe Text)

The Amazon Resource Name (ARN) of the Capacity Reservation.

capacityReservation_instanceMatchCriteria :: Lens' CapacityReservation (Maybe InstanceMatchCriteria)

Indicates the type of instance launches that the Capacity Reservation accepts. The options include:

open - The Capacity Reservation accepts all instances that have matching attributes (instance type, platform, and Availability Zone). Instances that have matching attributes launch into the Capacity Reservation automatically without specifying any additional parameters.

targeted - The Capacity Reservation only accepts instances that have matching attributes (instance type, platform, and Availability Zone), and explicitly target the Capacity Reservation. This ensures that only permitted instances can use the reserved capacity.

capacityReservation_capacityReservationId :: Lens' CapacityReservation (Maybe Text)

The ID of the Capacity Reservation.

capacityReservation_instancePlatform :: Lens' CapacityReservation (Maybe CapacityReservationInstancePlatform)

The type of operating system for which the Capacity Reservation reserves capacity.

capacityReservation_ephemeralStorage :: Lens' CapacityReservation (Maybe Bool)

Indicates whether the Capacity Reservation supports instances with temporary, block-level storage.

capacityReservation_endDate :: Lens' CapacityReservation (Maybe UTCTime)

The date and time at which the Capacity Reservation expires. When a Capacity Reservation expires, the reserved capacity is released and you can no longer launch instances into it. The Capacity Reservation's state changes to expired when it reaches its end date and time.

capacityReservation_totalInstanceCount :: Lens' CapacityReservation (Maybe Int)

The total number of instances for which the Capacity Reservation reserves capacity.