Network.AWS.EC2.Types.CapacityReservationSpecification

Description

Documentation

data CapacityReservationSpecification

Describes an instance's Capacity Reservation targeting option. You can specify only one parameter at a time. If you specify CapacityReservationPreference and CapacityReservationTarget, the request fails.

Use the CapacityReservationPreference parameter to configure the instance to run as an On-Demand Instance or to run in any open Capacity Reservation that has matching attributes (instance type, platform, Availability Zone). Use the CapacityReservationTarget parameter to explicitly target a specific Capacity Reservation or a Capacity Reservation group.

See: newCapacityReservationSpecification smart constructor.

Constructors

CapacityReservationSpecification'

Fields

capacityReservationTarget :: Maybe CapacityReservationTarget

Information about the target Capacity Reservation or Capacity Reservation group.

capacityReservationPreference :: Maybe CapacityReservationPreference

Indicates the instance's Capacity Reservation preferences. Possible preferences include:

open - The instance can run in any open Capacity Reservation that has matching attributes (instance type, platform, Availability Zone).

none - The instance avoids running in a Capacity Reservation even if one is available. The instance runs as an On-Demand Instance.

newCapacityReservationSpecification :: CapacityReservationSpecification

Create a value of CapacityReservationSpecification with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:capacityReservationTarget:CapacityReservationSpecification', capacityReservationSpecification_capacityReservationTarget - Information about the target Capacity Reservation or Capacity Reservation group.

$sel:capacityReservationPreference:CapacityReservationSpecification', capacityReservationSpecification_capacityReservationPreference - Indicates the instance's Capacity Reservation preferences. Possible preferences include:

open - The instance can run in any open Capacity Reservation that has matching attributes (instance type, platform, Availability Zone).

none - The instance avoids running in a Capacity Reservation even if one is available. The instance runs as an On-Demand Instance.

capacityReservationSpecification_capacityReservationTarget :: Lens' CapacityReservationSpecification (Maybe CapacityReservationTarget)

Information about the target Capacity Reservation or Capacity Reservation group.

capacityReservationSpecification_capacityReservationPreference :: Lens' CapacityReservationSpecification (Maybe CapacityReservationPreference)

Indicates the instance's Capacity Reservation preferences. Possible preferences include:

open - The instance can run in any open Capacity Reservation that has matching attributes (instance type, platform, Availability Zone).

none - The instance avoids running in a Capacity Reservation even if one is available. The instance runs as an On-Demand Instance.