Network.AWS.EC2.Types.ClientVpnAuthenticationRequest

Description

Documentation

data ClientVpnAuthenticationRequest

Describes the authentication method to be used by a Client VPN endpoint. For more information, see Authentication in the Client VPN Administrator Guide.

See: newClientVpnAuthenticationRequest smart constructor.

Constructors

ClientVpnAuthenticationRequest'

Fields

federatedAuthentication :: Maybe FederatedAuthenticationRequest

Information about the IAM SAML identity provider to be used, if applicable. You must provide this information if Type is federated-authentication.

activeDirectory :: Maybe DirectoryServiceAuthenticationRequest

Information about the Active Directory to be used, if applicable. You must provide this information if Type is directory-service-authentication.

mutualAuthentication :: Maybe CertificateAuthenticationRequest

Information about the authentication certificates to be used, if applicable. You must provide this information if Type is certificate-authentication.

type' :: Maybe ClientVpnAuthenticationType

The type of client authentication to be used.

newClientVpnAuthenticationRequest :: ClientVpnAuthenticationRequest

Create a value of ClientVpnAuthenticationRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:federatedAuthentication:ClientVpnAuthenticationRequest', clientVpnAuthenticationRequest_federatedAuthentication - Information about the IAM SAML identity provider to be used, if applicable. You must provide this information if Type is federated-authentication.

$sel:activeDirectory:ClientVpnAuthenticationRequest', clientVpnAuthenticationRequest_activeDirectory - Information about the Active Directory to be used, if applicable. You must provide this information if Type is directory-service-authentication.

$sel:mutualAuthentication:ClientVpnAuthenticationRequest', clientVpnAuthenticationRequest_mutualAuthentication - Information about the authentication certificates to be used, if applicable. You must provide this information if Type is certificate-authentication.

$sel:type':ClientVpnAuthenticationRequest', clientVpnAuthenticationRequest_type - The type of client authentication to be used.

clientVpnAuthenticationRequest_federatedAuthentication :: Lens' ClientVpnAuthenticationRequest (Maybe FederatedAuthenticationRequest)

Information about the IAM SAML identity provider to be used, if applicable. You must provide this information if Type is federated-authentication.

clientVpnAuthenticationRequest_activeDirectory :: Lens' ClientVpnAuthenticationRequest (Maybe DirectoryServiceAuthenticationRequest)

Information about the Active Directory to be used, if applicable. You must provide this information if Type is directory-service-authentication.

clientVpnAuthenticationRequest_mutualAuthentication :: Lens' ClientVpnAuthenticationRequest (Maybe CertificateAuthenticationRequest)

Information about the authentication certificates to be used, if applicable. You must provide this information if Type is certificate-authentication.

clientVpnAuthenticationRequest_type :: Lens' ClientVpnAuthenticationRequest (Maybe ClientVpnAuthenticationType)

The type of client authentication to be used.