Network.AWS.EC2.Types.ClientVpnConnection

Description

Documentation

data ClientVpnConnection

Describes a client connection.

See: newClientVpnConnection smart constructor.

Constructors

ClientVpnConnection'

Fields

clientVpnEndpointId :: Maybe Text

The ID of the Client VPN endpoint to which the client is connected.

status :: Maybe ClientVpnConnectionStatus

The current state of the client connection.

egressBytes :: Maybe Text

The number of bytes received by the client.

ingressPackets :: Maybe Text

The number of packets sent by the client.

connectionId :: Maybe Text

The ID of the client connection.

connectionEstablishedTime :: Maybe Text

The date and time the client connection was established.

postureComplianceStatuses :: Maybe [Text]

The statuses returned by the client connect handler for posture compliance, if applicable.

commonName :: Maybe Text

The common name associated with the client. This is either the name of the client certificate, or the Active Directory user name.

connectionEndTime :: Maybe Text

The date and time the client connection was terminated.

timestamp :: Maybe Text

The current date and time.

ingressBytes :: Maybe Text

The number of bytes sent by the client.

egressPackets :: Maybe Text

The number of packets received by the client.

username :: Maybe Text

The username of the client who established the client connection. This information is only provided if Active Directory client authentication is used.

clientIp :: Maybe Text

The IP address of the client.

newClientVpnConnection :: ClientVpnConnection

Create a value of ClientVpnConnection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientVpnEndpointId:ClientVpnConnection', clientVpnConnection_clientVpnEndpointId - The ID of the Client VPN endpoint to which the client is connected.

$sel:status:ClientVpnConnection', clientVpnConnection_status - The current state of the client connection.

$sel:egressBytes:ClientVpnConnection', clientVpnConnection_egressBytes - The number of bytes received by the client.

$sel:ingressPackets:ClientVpnConnection', clientVpnConnection_ingressPackets - The number of packets sent by the client.

$sel:connectionId:ClientVpnConnection', clientVpnConnection_connectionId - The ID of the client connection.

$sel:connectionEstablishedTime:ClientVpnConnection', clientVpnConnection_connectionEstablishedTime - The date and time the client connection was established.

$sel:postureComplianceStatuses:ClientVpnConnection', clientVpnConnection_postureComplianceStatuses - The statuses returned by the client connect handler for posture compliance, if applicable.

$sel:commonName:ClientVpnConnection', clientVpnConnection_commonName - The common name associated with the client. This is either the name of the client certificate, or the Active Directory user name.

$sel:connectionEndTime:ClientVpnConnection', clientVpnConnection_connectionEndTime - The date and time the client connection was terminated.

$sel:timestamp:ClientVpnConnection', clientVpnConnection_timestamp - The current date and time.

$sel:ingressBytes:ClientVpnConnection', clientVpnConnection_ingressBytes - The number of bytes sent by the client.

$sel:egressPackets:ClientVpnConnection', clientVpnConnection_egressPackets - The number of packets received by the client.

$sel:username:ClientVpnConnection', clientVpnConnection_username - The username of the client who established the client connection. This information is only provided if Active Directory client authentication is used.

$sel:clientIp:ClientVpnConnection', clientVpnConnection_clientIp - The IP address of the client.

clientVpnConnection_clientVpnEndpointId :: Lens' ClientVpnConnection (Maybe Text)

The ID of the Client VPN endpoint to which the client is connected.

clientVpnConnection_status :: Lens' ClientVpnConnection (Maybe ClientVpnConnectionStatus)

The current state of the client connection.

clientVpnConnection_egressBytes :: Lens' ClientVpnConnection (Maybe Text)

The number of bytes received by the client.

clientVpnConnection_ingressPackets :: Lens' ClientVpnConnection (Maybe Text)

The number of packets sent by the client.

clientVpnConnection_connectionId :: Lens' ClientVpnConnection (Maybe Text)

The ID of the client connection.

clientVpnConnection_connectionEstablishedTime :: Lens' ClientVpnConnection (Maybe Text)

The date and time the client connection was established.

clientVpnConnection_postureComplianceStatuses :: Lens' ClientVpnConnection (Maybe [Text])

The statuses returned by the client connect handler for posture compliance, if applicable.

clientVpnConnection_commonName :: Lens' ClientVpnConnection (Maybe Text)

The common name associated with the client. This is either the name of the client certificate, or the Active Directory user name.

clientVpnConnection_connectionEndTime :: Lens' ClientVpnConnection (Maybe Text)

The date and time the client connection was terminated.

clientVpnConnection_timestamp :: Lens' ClientVpnConnection (Maybe Text)

The current date and time.

clientVpnConnection_ingressBytes :: Lens' ClientVpnConnection (Maybe Text)

The number of bytes sent by the client.

clientVpnConnection_egressPackets :: Lens' ClientVpnConnection (Maybe Text)

The number of packets received by the client.

clientVpnConnection_username :: Lens' ClientVpnConnection (Maybe Text)

The username of the client who established the client connection. This information is only provided if Active Directory client authentication is used.

clientVpnConnection_clientIp :: Lens' ClientVpnConnection (Maybe Text)

The IP address of the client.