Network.AWS.EC2.Types.ClientVpnEndpoint

Description

Documentation

data ClientVpnEndpoint

Describes a Client VPN endpoint.

See: newClientVpnEndpoint smart constructor.

Constructors

ClientVpnEndpoint'

Fields

creationTime :: Maybe Text

The date and time the Client VPN endpoint was created.

status :: Maybe ClientVpnEndpointStatus

The current state of the Client VPN endpoint.

associatedTargetNetworks :: Maybe [AssociatedTargetNetwork]

Information about the associated target networks. A target network is a subnet in a VPC.

securityGroupIds :: Maybe [Text]

The IDs of the security groups for the target network.

connectionLogOptions :: Maybe ConnectionLogResponseOptions

Information about the client connection logging options for the Client VPN endpoint.

splitTunnel :: Maybe Bool

Indicates whether split-tunnel is enabled in the Client VPN endpoint.

For information about split-tunnel VPN endpoints, see Split-Tunnel Client VPN endpoint in the Client VPN Administrator Guide.

transportProtocol :: Maybe TransportProtocol

The transport protocol used by the Client VPN endpoint.

vpcId :: Maybe Text

The ID of the VPC.

vpnPort :: Maybe Int

The port number for the Client VPN endpoint.

deletionTime :: Maybe Text

The date and time the Client VPN endpoint was deleted, if applicable.

clientCidrBlock :: Maybe Text

The IPv4 address range, in CIDR notation, from which client IP addresses are assigned.

dnsServers :: Maybe [Text]

Information about the DNS servers to be used for DNS resolution.

clientVpnEndpointId :: Maybe Text

The ID of the Client VPN endpoint.

clientConnectOptions :: Maybe ClientConnectResponseOptions

The options for managing connection authorization for new client connections.

serverCertificateArn :: Maybe Text

The ARN of the server certificate.

authenticationOptions :: Maybe [ClientVpnAuthentication]

Information about the authentication method used by the Client VPN endpoint.

selfServicePortalUrl :: Maybe Text

The URL of the self-service portal.

description :: Maybe Text

A brief description of the endpoint.

dnsName :: Maybe Text

The DNS name to be used by clients when connecting to the Client VPN endpoint.

vpnProtocol :: Maybe VpnProtocol

The protocol used by the VPN session.

tags :: Maybe [Tag]

Any tags assigned to the Client VPN endpoint.

newClientVpnEndpoint :: ClientVpnEndpoint

Create a value of ClientVpnEndpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:ClientVpnEndpoint', clientVpnEndpoint_creationTime - The date and time the Client VPN endpoint was created.

$sel:status:ClientVpnEndpoint', clientVpnEndpoint_status - The current state of the Client VPN endpoint.

$sel:associatedTargetNetworks:ClientVpnEndpoint', clientVpnEndpoint_associatedTargetNetworks - Information about the associated target networks. A target network is a subnet in a VPC.

$sel:securityGroupIds:ClientVpnEndpoint', clientVpnEndpoint_securityGroupIds - The IDs of the security groups for the target network.

$sel:connectionLogOptions:ClientVpnEndpoint', clientVpnEndpoint_connectionLogOptions - Information about the client connection logging options for the Client VPN endpoint.

$sel:splitTunnel:ClientVpnEndpoint', clientVpnEndpoint_splitTunnel - Indicates whether split-tunnel is enabled in the Client VPN endpoint.

For information about split-tunnel VPN endpoints, see Split-Tunnel Client VPN endpoint in the Client VPN Administrator Guide.

$sel:transportProtocol:ClientVpnEndpoint', clientVpnEndpoint_transportProtocol - The transport protocol used by the Client VPN endpoint.

$sel:vpcId:ClientVpnEndpoint', clientVpnEndpoint_vpcId - The ID of the VPC.

$sel:vpnPort:ClientVpnEndpoint', clientVpnEndpoint_vpnPort - The port number for the Client VPN endpoint.

$sel:deletionTime:ClientVpnEndpoint', clientVpnEndpoint_deletionTime - The date and time the Client VPN endpoint was deleted, if applicable.

$sel:clientCidrBlock:ClientVpnEndpoint', clientVpnEndpoint_clientCidrBlock - The IPv4 address range, in CIDR notation, from which client IP addresses are assigned.

$sel:dnsServers:ClientVpnEndpoint', clientVpnEndpoint_dnsServers - Information about the DNS servers to be used for DNS resolution.

$sel:clientVpnEndpointId:ClientVpnEndpoint', clientVpnEndpoint_clientVpnEndpointId - The ID of the Client VPN endpoint.

$sel:clientConnectOptions:ClientVpnEndpoint', clientVpnEndpoint_clientConnectOptions - The options for managing connection authorization for new client connections.

$sel:serverCertificateArn:ClientVpnEndpoint', clientVpnEndpoint_serverCertificateArn - The ARN of the server certificate.

$sel:authenticationOptions:ClientVpnEndpoint', clientVpnEndpoint_authenticationOptions - Information about the authentication method used by the Client VPN endpoint.

$sel:selfServicePortalUrl:ClientVpnEndpoint', clientVpnEndpoint_selfServicePortalUrl - The URL of the self-service portal.

$sel:description:ClientVpnEndpoint', clientVpnEndpoint_description - A brief description of the endpoint.

$sel:dnsName:ClientVpnEndpoint', clientVpnEndpoint_dnsName - The DNS name to be used by clients when connecting to the Client VPN endpoint.

$sel:vpnProtocol:ClientVpnEndpoint', clientVpnEndpoint_vpnProtocol - The protocol used by the VPN session.

$sel:tags:ClientVpnEndpoint', clientVpnEndpoint_tags - Any tags assigned to the Client VPN endpoint.

clientVpnEndpoint_creationTime :: Lens' ClientVpnEndpoint (Maybe Text)

The date and time the Client VPN endpoint was created.

clientVpnEndpoint_status :: Lens' ClientVpnEndpoint (Maybe ClientVpnEndpointStatus)

The current state of the Client VPN endpoint.

clientVpnEndpoint_associatedTargetNetworks :: Lens' ClientVpnEndpoint (Maybe [AssociatedTargetNetwork])

Information about the associated target networks. A target network is a subnet in a VPC.

clientVpnEndpoint_securityGroupIds :: Lens' ClientVpnEndpoint (Maybe [Text])

The IDs of the security groups for the target network.

clientVpnEndpoint_connectionLogOptions :: Lens' ClientVpnEndpoint (Maybe ConnectionLogResponseOptions)

Information about the client connection logging options for the Client VPN endpoint.

clientVpnEndpoint_splitTunnel :: Lens' ClientVpnEndpoint (Maybe Bool)

Indicates whether split-tunnel is enabled in the Client VPN endpoint.

For information about split-tunnel VPN endpoints, see Split-Tunnel Client VPN endpoint in the Client VPN Administrator Guide.

clientVpnEndpoint_transportProtocol :: Lens' ClientVpnEndpoint (Maybe TransportProtocol)

The transport protocol used by the Client VPN endpoint.

clientVpnEndpoint_vpcId :: Lens' ClientVpnEndpoint (Maybe Text)

The ID of the VPC.

clientVpnEndpoint_vpnPort :: Lens' ClientVpnEndpoint (Maybe Int)

The port number for the Client VPN endpoint.

clientVpnEndpoint_deletionTime :: Lens' ClientVpnEndpoint (Maybe Text)

The date and time the Client VPN endpoint was deleted, if applicable.

clientVpnEndpoint_clientCidrBlock :: Lens' ClientVpnEndpoint (Maybe Text)

The IPv4 address range, in CIDR notation, from which client IP addresses are assigned.

clientVpnEndpoint_dnsServers :: Lens' ClientVpnEndpoint (Maybe [Text])

Information about the DNS servers to be used for DNS resolution.

clientVpnEndpoint_clientVpnEndpointId :: Lens' ClientVpnEndpoint (Maybe Text)

The ID of the Client VPN endpoint.

clientVpnEndpoint_clientConnectOptions :: Lens' ClientVpnEndpoint (Maybe ClientConnectResponseOptions)

The options for managing connection authorization for new client connections.

clientVpnEndpoint_serverCertificateArn :: Lens' ClientVpnEndpoint (Maybe Text)

The ARN of the server certificate.

clientVpnEndpoint_authenticationOptions :: Lens' ClientVpnEndpoint (Maybe [ClientVpnAuthentication])

Information about the authentication method used by the Client VPN endpoint.

clientVpnEndpoint_selfServicePortalUrl :: Lens' ClientVpnEndpoint (Maybe Text)

The URL of the self-service portal.

clientVpnEndpoint_description :: Lens' ClientVpnEndpoint (Maybe Text)

A brief description of the endpoint.

clientVpnEndpoint_dnsName :: Lens' ClientVpnEndpoint (Maybe Text)

The DNS name to be used by clients when connecting to the Client VPN endpoint.

clientVpnEndpoint_vpnProtocol :: Lens' ClientVpnEndpoint (Maybe VpnProtocol)

The protocol used by the VPN session.

clientVpnEndpoint_tags :: Lens' ClientVpnEndpoint (Maybe [Tag])

Any tags assigned to the Client VPN endpoint.