Network.AWS.EC2.Types.ClientVpnRoute

Description

Documentation

data ClientVpnRoute

Information about a Client VPN endpoint route.

See: newClientVpnRoute smart constructor.

Constructors

ClientVpnRoute'

Fields

status :: Maybe ClientVpnRouteStatus

The current state of the route.

origin :: Maybe Text

Indicates how the route was associated with the Client VPN endpoint. associate indicates that the route was automatically added when the target network was associated with the Client VPN endpoint. add-route indicates that the route was manually added using the CreateClientVpnRoute action.

clientVpnEndpointId :: Maybe Text

The ID of the Client VPN endpoint with which the route is associated.

targetSubnet :: Maybe Text

The ID of the subnet through which traffic is routed.

destinationCidr :: Maybe Text

The IPv4 address range, in CIDR notation, of the route destination.

type' :: Maybe Text

The route type.

description :: Maybe Text

A brief description of the route.

newClientVpnRoute :: ClientVpnRoute

Create a value of ClientVpnRoute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ClientVpnRoute', clientVpnRoute_status - The current state of the route.

$sel:origin:ClientVpnRoute', clientVpnRoute_origin - Indicates how the route was associated with the Client VPN endpoint. associate indicates that the route was automatically added when the target network was associated with the Client VPN endpoint. add-route indicates that the route was manually added using the CreateClientVpnRoute action.

$sel:clientVpnEndpointId:ClientVpnRoute', clientVpnRoute_clientVpnEndpointId - The ID of the Client VPN endpoint with which the route is associated.

$sel:targetSubnet:ClientVpnRoute', clientVpnRoute_targetSubnet - The ID of the subnet through which traffic is routed.

$sel:destinationCidr:ClientVpnRoute', clientVpnRoute_destinationCidr - The IPv4 address range, in CIDR notation, of the route destination.

$sel:type':ClientVpnRoute', clientVpnRoute_type - The route type.

$sel:description:ClientVpnRoute', clientVpnRoute_description - A brief description of the route.

clientVpnRoute_status :: Lens' ClientVpnRoute (Maybe ClientVpnRouteStatus)

The current state of the route.

clientVpnRoute_origin :: Lens' ClientVpnRoute (Maybe Text)

Indicates how the route was associated with the Client VPN endpoint. associate indicates that the route was automatically added when the target network was associated with the Client VPN endpoint. add-route indicates that the route was manually added using the CreateClientVpnRoute action.

clientVpnRoute_clientVpnEndpointId :: Lens' ClientVpnRoute (Maybe Text)

The ID of the Client VPN endpoint with which the route is associated.

clientVpnRoute_targetSubnet :: Lens' ClientVpnRoute (Maybe Text)

The ID of the subnet through which traffic is routed.

clientVpnRoute_destinationCidr :: Lens' ClientVpnRoute (Maybe Text)

The IPv4 address range, in CIDR notation, of the route destination.

clientVpnRoute_type :: Lens' ClientVpnRoute (Maybe Text)

The route type.

clientVpnRoute_description :: Lens' ClientVpnRoute (Maybe Text)

A brief description of the route.