Network.AWS.EC2.Types.ConnectionLogResponseOptions

Description

Documentation

data ConnectionLogResponseOptions

Information about the client connection logging options for a Client VPN endpoint.

See: newConnectionLogResponseOptions smart constructor.

Constructors

ConnectionLogResponseOptions'

Fields

cloudwatchLogStream :: Maybe Text

The name of the Amazon CloudWatch Logs log stream to which connection logging data is published.

enabled :: Maybe Bool

Indicates whether client connection logging is enabled for the Client VPN endpoint.

cloudwatchLogGroup :: Maybe Text

The name of the Amazon CloudWatch Logs log group to which connection logging data is published.

newConnectionLogResponseOptions :: ConnectionLogResponseOptions

Create a value of ConnectionLogResponseOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cloudwatchLogStream:ConnectionLogResponseOptions', connectionLogResponseOptions_cloudwatchLogStream - The name of the Amazon CloudWatch Logs log stream to which connection logging data is published.

$sel:enabled:ConnectionLogResponseOptions', connectionLogResponseOptions_enabled - Indicates whether client connection logging is enabled for the Client VPN endpoint.

$sel:cloudwatchLogGroup:ConnectionLogResponseOptions', connectionLogResponseOptions_cloudwatchLogGroup - The name of the Amazon CloudWatch Logs log group to which connection logging data is published.

connectionLogResponseOptions_cloudwatchLogStream :: Lens' ConnectionLogResponseOptions (Maybe Text)

The name of the Amazon CloudWatch Logs log stream to which connection logging data is published.

connectionLogResponseOptions_enabled :: Lens' ConnectionLogResponseOptions (Maybe Bool)

Indicates whether client connection logging is enabled for the Client VPN endpoint.

connectionLogResponseOptions_cloudwatchLogGroup :: Lens' ConnectionLogResponseOptions (Maybe Text)

The name of the Amazon CloudWatch Logs log group to which connection logging data is published.