Network.AWS.EC2.Types.ConversionTask

Description

Documentation

data ConversionTask

Describes a conversion task.

See: newConversionTask smart constructor.

Constructors

ConversionTask'

Fields

importInstance :: Maybe ImportInstanceTaskDetails

If the task is for importing an instance, this contains information about the import instance task.

state :: Maybe ConversionTaskState

The state of the conversion task.

statusMessage :: Maybe Text

The status message related to the conversion task.

importVolume :: Maybe ImportVolumeTaskDetails

If the task is for importing a volume, this contains information about the import volume task.

conversionTaskId :: Maybe Text

The ID of the conversion task.

expirationTime :: Maybe Text

The time when the task expires. If the upload isn't complete before the expiration time, we automatically cancel the task.

tags :: Maybe [Tag]

Any tags assigned to the task.

newConversionTask :: ConversionTask

Create a value of ConversionTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:importInstance:ConversionTask', conversionTask_importInstance - If the task is for importing an instance, this contains information about the import instance task.

$sel:state:ConversionTask', conversionTask_state - The state of the conversion task.

$sel:statusMessage:ConversionTask', conversionTask_statusMessage - The status message related to the conversion task.

$sel:importVolume:ConversionTask', conversionTask_importVolume - If the task is for importing a volume, this contains information about the import volume task.

$sel:conversionTaskId:ConversionTask', conversionTask_conversionTaskId - The ID of the conversion task.

$sel:expirationTime:ConversionTask', conversionTask_expirationTime - The time when the task expires. If the upload isn't complete before the expiration time, we automatically cancel the task.

$sel:tags:ConversionTask', conversionTask_tags - Any tags assigned to the task.

conversionTask_importInstance :: Lens' ConversionTask (Maybe ImportInstanceTaskDetails)

If the task is for importing an instance, this contains information about the import instance task.

conversionTask_state :: Lens' ConversionTask (Maybe ConversionTaskState)

The state of the conversion task.

conversionTask_statusMessage :: Lens' ConversionTask (Maybe Text)

The status message related to the conversion task.

conversionTask_importVolume :: Lens' ConversionTask (Maybe ImportVolumeTaskDetails)

If the task is for importing a volume, this contains information about the import volume task.

conversionTask_conversionTaskId :: Lens' ConversionTask (Maybe Text)

The ID of the conversion task.

conversionTask_expirationTime :: Lens' ConversionTask (Maybe Text)

The time when the task expires. If the upload isn't complete before the expiration time, we automatically cancel the task.

conversionTask_tags :: Lens' ConversionTask (Maybe [Tag])

Any tags assigned to the task.