Network.AWS.EC2.Types.CreateFleetInstance

Description

Documentation

data CreateFleetInstance

Describes the instances that were launched by the fleet.

See: newCreateFleetInstance smart constructor.

Constructors

CreateFleetInstance'

Fields

instanceIds :: Maybe [Text]

The IDs of the instances.

platform :: Maybe PlatformValues

The value is Windows for Windows instances. Otherwise, the value is blank.

launchTemplateAndOverrides :: Maybe LaunchTemplateAndOverridesResponse

The launch templates and overrides that were used for launching the instances. The values that you specify in the Overrides replace the values in the launch template.

instanceType :: Maybe InstanceType

The instance type.

lifecycle :: Maybe InstanceLifecycle

Indicates if the instance that was launched is a Spot Instance or On-Demand Instance.

newCreateFleetInstance :: CreateFleetInstance

Create a value of CreateFleetInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceIds:CreateFleetInstance', createFleetInstance_instanceIds - The IDs of the instances.

$sel:platform:CreateFleetInstance', createFleetInstance_platform - The value is Windows for Windows instances. Otherwise, the value is blank.

$sel:launchTemplateAndOverrides:CreateFleetInstance', createFleetInstance_launchTemplateAndOverrides - The launch templates and overrides that were used for launching the instances. The values that you specify in the Overrides replace the values in the launch template.

$sel:instanceType:CreateFleetInstance', createFleetInstance_instanceType - The instance type.

$sel:lifecycle:CreateFleetInstance', createFleetInstance_lifecycle - Indicates if the instance that was launched is a Spot Instance or On-Demand Instance.

createFleetInstance_instanceIds :: Lens' CreateFleetInstance (Maybe [Text])

The IDs of the instances.

createFleetInstance_platform :: Lens' CreateFleetInstance (Maybe PlatformValues)

The value is Windows for Windows instances. Otherwise, the value is blank.

createFleetInstance_launchTemplateAndOverrides :: Lens' CreateFleetInstance (Maybe LaunchTemplateAndOverridesResponse)

The launch templates and overrides that were used for launching the instances. The values that you specify in the Overrides replace the values in the launch template.

createFleetInstance_instanceType :: Lens' CreateFleetInstance (Maybe InstanceType)

The instance type.

createFleetInstance_lifecycle :: Lens' CreateFleetInstance (Maybe InstanceLifecycle)

Indicates if the instance that was launched is a Spot Instance or On-Demand Instance.