Network.AWS.EC2.Types.CreateTransitGatewayVpcAttachmentRequestOptions

Description

Documentation

data CreateTransitGatewayVpcAttachmentRequestOptions

Describes the options for a VPC attachment.

See: newCreateTransitGatewayVpcAttachmentRequestOptions smart constructor.

Constructors

CreateTransitGatewayVpcAttachmentRequestOptions'

Fields

applianceModeSupport :: Maybe ApplianceModeSupportValue

Enable or disable support for appliance mode. If enabled, a traffic flow between a source and destination uses the same Availability Zone for the VPC attachment for the lifetime of that flow. The default is disable.

ipv6Support :: Maybe Ipv6SupportValue

Enable or disable IPv6 support. The default is disable.

dnsSupport :: Maybe DnsSupportValue

Enable or disable DNS support. The default is enable.

newCreateTransitGatewayVpcAttachmentRequestOptions :: CreateTransitGatewayVpcAttachmentRequestOptions

Create a value of CreateTransitGatewayVpcAttachmentRequestOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applianceModeSupport:CreateTransitGatewayVpcAttachmentRequestOptions', createTransitGatewayVpcAttachmentRequestOptions_applianceModeSupport - Enable or disable support for appliance mode. If enabled, a traffic flow between a source and destination uses the same Availability Zone for the VPC attachment for the lifetime of that flow. The default is disable.

$sel:ipv6Support:CreateTransitGatewayVpcAttachmentRequestOptions', createTransitGatewayVpcAttachmentRequestOptions_ipv6Support - Enable or disable IPv6 support. The default is disable.

$sel:dnsSupport:CreateTransitGatewayVpcAttachmentRequestOptions', createTransitGatewayVpcAttachmentRequestOptions_dnsSupport - Enable or disable DNS support. The default is enable.

createTransitGatewayVpcAttachmentRequestOptions_applianceModeSupport :: Lens' CreateTransitGatewayVpcAttachmentRequestOptions (Maybe ApplianceModeSupportValue)

Enable or disable support for appliance mode. If enabled, a traffic flow between a source and destination uses the same Availability Zone for the VPC attachment for the lifetime of that flow. The default is disable.

createTransitGatewayVpcAttachmentRequestOptions_ipv6Support :: Lens' CreateTransitGatewayVpcAttachmentRequestOptions (Maybe Ipv6SupportValue)

Enable or disable IPv6 support. The default is disable.

createTransitGatewayVpcAttachmentRequestOptions_dnsSupport :: Lens' CreateTransitGatewayVpcAttachmentRequestOptions (Maybe DnsSupportValue)

Enable or disable DNS support. The default is enable.