Network.AWS.EC2.Types.DescribeFastSnapshotRestoreSuccessItem

Description

Documentation

data DescribeFastSnapshotRestoreSuccessItem

Describes fast snapshot restores for a snapshot.

See: newDescribeFastSnapshotRestoreSuccessItem smart constructor.

Constructors

DescribeFastSnapshotRestoreSuccessItem'

Fields

ownerAlias :: Maybe Text

The Amazon Web Services owner alias that enabled fast snapshot restores on the snapshot. This is intended for future use.

stateTransitionReason :: Maybe Text

The reason for the state transition. The possible values are as follows:

Client.UserInitiated - The state successfully transitioned to enabling or disabling.

Client.UserInitiated - Lifecycle state transition - The state successfully transitioned to optimizing, enabled, or disabled.

ownerId :: Maybe Text

The ID of the Amazon Web Services account that enabled fast snapshot restores on the snapshot.

optimizingTime :: Maybe ISO8601

The time at which fast snapshot restores entered the optimizing state.

availabilityZone :: Maybe Text

The Availability Zone.

state :: Maybe FastSnapshotRestoreStateCode

The state of fast snapshot restores.

disablingTime :: Maybe ISO8601

The time at which fast snapshot restores entered the disabling state.

snapshotId :: Maybe Text

The ID of the snapshot.

enablingTime :: Maybe ISO8601

The time at which fast snapshot restores entered the enabling state.

enabledTime :: Maybe ISO8601

The time at which fast snapshot restores entered the enabled state.

disabledTime :: Maybe ISO8601

The time at which fast snapshot restores entered the disabled state.

newDescribeFastSnapshotRestoreSuccessItem :: DescribeFastSnapshotRestoreSuccessItem

Create a value of DescribeFastSnapshotRestoreSuccessItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ownerAlias:DescribeFastSnapshotRestoreSuccessItem', describeFastSnapshotRestoreSuccessItem_ownerAlias - The Amazon Web Services owner alias that enabled fast snapshot restores on the snapshot. This is intended for future use.

$sel:stateTransitionReason:DescribeFastSnapshotRestoreSuccessItem', describeFastSnapshotRestoreSuccessItem_stateTransitionReason - The reason for the state transition. The possible values are as follows:

Client.UserInitiated - The state successfully transitioned to enabling or disabling.

Client.UserInitiated - Lifecycle state transition - The state successfully transitioned to optimizing, enabled, or disabled.

$sel:ownerId:DescribeFastSnapshotRestoreSuccessItem', describeFastSnapshotRestoreSuccessItem_ownerId - The ID of the Amazon Web Services account that enabled fast snapshot restores on the snapshot.

$sel:optimizingTime:DescribeFastSnapshotRestoreSuccessItem', describeFastSnapshotRestoreSuccessItem_optimizingTime - The time at which fast snapshot restores entered the optimizing state.

$sel:availabilityZone:DescribeFastSnapshotRestoreSuccessItem', describeFastSnapshotRestoreSuccessItem_availabilityZone - The Availability Zone.

$sel:state:DescribeFastSnapshotRestoreSuccessItem', describeFastSnapshotRestoreSuccessItem_state - The state of fast snapshot restores.

$sel:disablingTime:DescribeFastSnapshotRestoreSuccessItem', describeFastSnapshotRestoreSuccessItem_disablingTime - The time at which fast snapshot restores entered the disabling state.

$sel:snapshotId:DescribeFastSnapshotRestoreSuccessItem', describeFastSnapshotRestoreSuccessItem_snapshotId - The ID of the snapshot.

$sel:enablingTime:DescribeFastSnapshotRestoreSuccessItem', describeFastSnapshotRestoreSuccessItem_enablingTime - The time at which fast snapshot restores entered the enabling state.

$sel:enabledTime:DescribeFastSnapshotRestoreSuccessItem', describeFastSnapshotRestoreSuccessItem_enabledTime - The time at which fast snapshot restores entered the enabled state.

$sel:disabledTime:DescribeFastSnapshotRestoreSuccessItem', describeFastSnapshotRestoreSuccessItem_disabledTime - The time at which fast snapshot restores entered the disabled state.

describeFastSnapshotRestoreSuccessItem_ownerAlias :: Lens' DescribeFastSnapshotRestoreSuccessItem (Maybe Text)

The Amazon Web Services owner alias that enabled fast snapshot restores on the snapshot. This is intended for future use.

describeFastSnapshotRestoreSuccessItem_stateTransitionReason :: Lens' DescribeFastSnapshotRestoreSuccessItem (Maybe Text)

The reason for the state transition. The possible values are as follows:

Client.UserInitiated - The state successfully transitioned to enabling or disabling.

Client.UserInitiated - Lifecycle state transition - The state successfully transitioned to optimizing, enabled, or disabled.

describeFastSnapshotRestoreSuccessItem_ownerId :: Lens' DescribeFastSnapshotRestoreSuccessItem (Maybe Text)

The ID of the Amazon Web Services account that enabled fast snapshot restores on the snapshot.

describeFastSnapshotRestoreSuccessItem_optimizingTime :: Lens' DescribeFastSnapshotRestoreSuccessItem (Maybe UTCTime)

The time at which fast snapshot restores entered the optimizing state.

describeFastSnapshotRestoreSuccessItem_availabilityZone :: Lens' DescribeFastSnapshotRestoreSuccessItem (Maybe Text)

The Availability Zone.

describeFastSnapshotRestoreSuccessItem_state :: Lens' DescribeFastSnapshotRestoreSuccessItem (Maybe FastSnapshotRestoreStateCode)

The state of fast snapshot restores.

describeFastSnapshotRestoreSuccessItem_disablingTime :: Lens' DescribeFastSnapshotRestoreSuccessItem (Maybe UTCTime)

The time at which fast snapshot restores entered the disabling state.

describeFastSnapshotRestoreSuccessItem_snapshotId :: Lens' DescribeFastSnapshotRestoreSuccessItem (Maybe Text)

The ID of the snapshot.

describeFastSnapshotRestoreSuccessItem_enablingTime :: Lens' DescribeFastSnapshotRestoreSuccessItem (Maybe UTCTime)

The time at which fast snapshot restores entered the enabling state.

describeFastSnapshotRestoreSuccessItem_enabledTime :: Lens' DescribeFastSnapshotRestoreSuccessItem (Maybe UTCTime)

The time at which fast snapshot restores entered the enabled state.

describeFastSnapshotRestoreSuccessItem_disabledTime :: Lens' DescribeFastSnapshotRestoreSuccessItem (Maybe UTCTime)

The time at which fast snapshot restores entered the disabled state.