Network.AWS.EC2.Types.DestinationOptionsResponse

Description

Documentation

data DestinationOptionsResponse

Describes the destination options for a flow log.

See: newDestinationOptionsResponse smart constructor.

Constructors

DestinationOptionsResponse'

Fields

perHourPartition :: Maybe Bool

Indicates whether to partition the flow log per hour.

fileFormat :: Maybe DestinationFileFormat

The format for the flow log.

hiveCompatiblePartitions :: Maybe Bool

Indicates whether to use Hive-compatible prefixes for flow logs stored in Amazon S3.

newDestinationOptionsResponse :: DestinationOptionsResponse

Create a value of DestinationOptionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:perHourPartition:DestinationOptionsResponse', destinationOptionsResponse_perHourPartition - Indicates whether to partition the flow log per hour.

$sel:fileFormat:DestinationOptionsResponse', destinationOptionsResponse_fileFormat - The format for the flow log.

$sel:hiveCompatiblePartitions:DestinationOptionsResponse', destinationOptionsResponse_hiveCompatiblePartitions - Indicates whether to use Hive-compatible prefixes for flow logs stored in Amazon S3.

destinationOptionsResponse_perHourPartition :: Lens' DestinationOptionsResponse (Maybe Bool)

Indicates whether to partition the flow log per hour.

destinationOptionsResponse_fileFormat :: Lens' DestinationOptionsResponse (Maybe DestinationFileFormat)

The format for the flow log.

destinationOptionsResponse_hiveCompatiblePartitions :: Lens' DestinationOptionsResponse (Maybe Bool)

Indicates whether to use Hive-compatible prefixes for flow logs stored in Amazon S3.