Network.AWS.EC2.Types.DiskImageDetail

Description

Documentation

data DiskImageDetail

Describes a disk image.

See: newDiskImageDetail smart constructor.

Constructors

DiskImageDetail'

Fields

bytes :: Integer

The size of the disk image, in GiB.

format :: DiskImageFormat

The disk image format.

importManifestUrl :: Text

A presigned URL for the import manifest stored in Amazon S3 and presented here as an Amazon S3 presigned URL. For information about creating a presigned URL for an Amazon S3 object, read the "Query String Request Authentication Alternative" section of the Authenticating REST Requests topic in the Amazon Simple Storage Service Developer Guide.

For information about the import manifest referenced by this API action, see VM Import Manifest.

newDiskImageDetail

Arguments

:: Integer
$sel:bytes:DiskImageDetail'

-> DiskImageFormat
$sel:format:DiskImageDetail'

-> Text
$sel:importManifestUrl:DiskImageDetail'

-> DiskImageDetail

Create a value of DiskImageDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bytes:DiskImageDetail', diskImageDetail_bytes - The size of the disk image, in GiB.

$sel:format:DiskImageDetail', diskImageDetail_format - The disk image format.

$sel:importManifestUrl:DiskImageDetail', diskImageDetail_importManifestUrl - A presigned URL for the import manifest stored in Amazon S3 and presented here as an Amazon S3 presigned URL. For information about creating a presigned URL for an Amazon S3 object, read the "Query String Request Authentication Alternative" section of the Authenticating REST Requests topic in the Amazon Simple Storage Service Developer Guide.

For information about the import manifest referenced by this API action, see VM Import Manifest.

diskImageDetail_bytes :: Lens' DiskImageDetail Integer

The size of the disk image, in GiB.

diskImageDetail_format :: Lens' DiskImageDetail DiskImageFormat

The disk image format.

diskImageDetail_importManifestUrl :: Lens' DiskImageDetail Text

A presigned URL for the import manifest stored in Amazon S3 and presented here as an Amazon S3 presigned URL. For information about creating a presigned URL for an Amazon S3 object, read the "Query String Request Authentication Alternative" section of the Authenticating REST Requests topic in the Amazon Simple Storage Service Developer Guide.

For information about the import manifest referenced by this API action, see VM Import Manifest.