Network.AWS.EC2.Types.EbsBlockDevice

Description

Documentation

data EbsBlockDevice

Describes a block device for an EBS volume.

See: newEbsBlockDevice smart constructor.

Constructors

EbsBlockDevice'

Fields

encrypted :: Maybe Bool

Indicates whether the encryption state of an EBS volume is changed while being restored from a backing snapshot. The effect of setting the encryption state to true depends on the volume origin (new or from a snapshot), starting encryption state, ownership, and whether encryption by default is enabled. For more information, see Amazon EBS encryption in the Amazon EC2 User Guide.

In no case can you remove encryption from an encrypted volume.

Encrypted volumes can only be attached to instances that support Amazon EBS encryption. For more information, see Supported instance types.

This parameter is not returned by .

outpostArn :: Maybe Text

The ARN of the Outpost on which the snapshot is stored.

throughput :: Maybe Int

The throughput that the volume supports, in MiB/s.

This parameter is valid only for gp3 volumes.

Valid Range: Minimum value of 125. Maximum value of 1000.

kmsKeyId :: Maybe Text

Identifier (key ID, key alias, ID ARN, or alias ARN) for a customer managed CMK under which the EBS volume is encrypted.

This parameter is only supported on BlockDeviceMapping objects called by RunInstances, RequestSpotFleet, and RequestSpotInstances.

deleteOnTermination :: Maybe Bool

Indicates whether the EBS volume is deleted on instance termination. For more information, see Preserving Amazon EBS volumes on instance termination in the Amazon EC2 User Guide.

snapshotId :: Maybe Text

The ID of the snapshot.

volumeType :: Maybe VolumeType

The volume type. For more information, see Amazon EBS volume types in the Amazon EC2 User Guide. If the volume type is io1 or io2, you must specify the IOPS that the volume supports.

iops :: Maybe Int

The number of I/O operations per second (IOPS). For gp3, io1, and io2 volumes, this represents the number of IOPS that are provisioned for the volume. For gp2 volumes, this represents the baseline performance of the volume and the rate at which the volume accumulates I/O credits for bursting.

The following are the supported values for each volume type:

gp3: 3,000-16,000 IOPS

io1: 100-64,000 IOPS

io2: 100-64,000 IOPS

For io1 and io2 volumes, we guarantee 64,000 IOPS only for Instances built on the Nitro System. Other instance families guarantee performance up to 32,000 IOPS.

This parameter is required for io1 and io2 volumes. The default for gp3 volumes is 3,000 IOPS. This parameter is not supported for gp2, st1, sc1, or standard volumes.

volumeSize :: Maybe Int

The size of the volume, in GiBs. You must specify either a snapshot ID or a volume size. If you specify a snapshot, the default is the snapshot size. You can specify a volume size that is equal to or larger than the snapshot size.

The following are the supported volumes sizes for each volume type:

gp2 and gp3:1-16,384

io1 and io2: 4-16,384

st1 and sc1: 125-16,384

standard: 1-1,024

newEbsBlockDevice :: EbsBlockDevice

Create a value of EbsBlockDevice with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:encrypted:EbsBlockDevice', ebsBlockDevice_encrypted - Indicates whether the encryption state of an EBS volume is changed while being restored from a backing snapshot. The effect of setting the encryption state to true depends on the volume origin (new or from a snapshot), starting encryption state, ownership, and whether encryption by default is enabled. For more information, see Amazon EBS encryption in the Amazon EC2 User Guide.

In no case can you remove encryption from an encrypted volume.

Encrypted volumes can only be attached to instances that support Amazon EBS encryption. For more information, see Supported instance types.

This parameter is not returned by .

$sel:outpostArn:EbsBlockDevice', ebsBlockDevice_outpostArn - The ARN of the Outpost on which the snapshot is stored.

$sel:throughput:EbsBlockDevice', ebsBlockDevice_throughput - The throughput that the volume supports, in MiB/s.

This parameter is valid only for gp3 volumes.

Valid Range: Minimum value of 125. Maximum value of 1000.

$sel:kmsKeyId:EbsBlockDevice', ebsBlockDevice_kmsKeyId - Identifier (key ID, key alias, ID ARN, or alias ARN) for a customer managed CMK under which the EBS volume is encrypted.

This parameter is only supported on BlockDeviceMapping objects called by RunInstances, RequestSpotFleet, and RequestSpotInstances.

$sel:deleteOnTermination:EbsBlockDevice', ebsBlockDevice_deleteOnTermination - Indicates whether the EBS volume is deleted on instance termination. For more information, see Preserving Amazon EBS volumes on instance termination in the Amazon EC2 User Guide.

$sel:snapshotId:EbsBlockDevice', ebsBlockDevice_snapshotId - The ID of the snapshot.

$sel:volumeType:EbsBlockDevice', ebsBlockDevice_volumeType - The volume type. For more information, see Amazon EBS volume types in the Amazon EC2 User Guide. If the volume type is io1 or io2, you must specify the IOPS that the volume supports.

$sel:iops:EbsBlockDevice', ebsBlockDevice_iops - The number of I/O operations per second (IOPS). For gp3, io1, and io2 volumes, this represents the number of IOPS that are provisioned for the volume. For gp2 volumes, this represents the baseline performance of the volume and the rate at which the volume accumulates I/O credits for bursting.

The following are the supported values for each volume type:

gp3: 3,000-16,000 IOPS

io1: 100-64,000 IOPS

io2: 100-64,000 IOPS

For io1 and io2 volumes, we guarantee 64,000 IOPS only for Instances built on the Nitro System. Other instance families guarantee performance up to 32,000 IOPS.

This parameter is required for io1 and io2 volumes. The default for gp3 volumes is 3,000 IOPS. This parameter is not supported for gp2, st1, sc1, or standard volumes.

$sel:volumeSize:EbsBlockDevice', ebsBlockDevice_volumeSize - The size of the volume, in GiBs. You must specify either a snapshot ID or a volume size. If you specify a snapshot, the default is the snapshot size. You can specify a volume size that is equal to or larger than the snapshot size.

The following are the supported volumes sizes for each volume type:

gp2 and gp3:1-16,384

io1 and io2: 4-16,384

st1 and sc1: 125-16,384

standard: 1-1,024

ebsBlockDevice_encrypted :: Lens' EbsBlockDevice (Maybe Bool)

Indicates whether the encryption state of an EBS volume is changed while being restored from a backing snapshot. The effect of setting the encryption state to true depends on the volume origin (new or from a snapshot), starting encryption state, ownership, and whether encryption by default is enabled. For more information, see Amazon EBS encryption in the Amazon EC2 User Guide.

In no case can you remove encryption from an encrypted volume.

Encrypted volumes can only be attached to instances that support Amazon EBS encryption. For more information, see Supported instance types.

This parameter is not returned by .

ebsBlockDevice_outpostArn :: Lens' EbsBlockDevice (Maybe Text)

The ARN of the Outpost on which the snapshot is stored.

ebsBlockDevice_throughput :: Lens' EbsBlockDevice (Maybe Int)

The throughput that the volume supports, in MiB/s.

This parameter is valid only for gp3 volumes.

Valid Range: Minimum value of 125. Maximum value of 1000.

ebsBlockDevice_kmsKeyId :: Lens' EbsBlockDevice (Maybe Text)

Identifier (key ID, key alias, ID ARN, or alias ARN) for a customer managed CMK under which the EBS volume is encrypted.

This parameter is only supported on BlockDeviceMapping objects called by RunInstances, RequestSpotFleet, and RequestSpotInstances.

ebsBlockDevice_deleteOnTermination :: Lens' EbsBlockDevice (Maybe Bool)

Indicates whether the EBS volume is deleted on instance termination. For more information, see Preserving Amazon EBS volumes on instance termination in the Amazon EC2 User Guide.

ebsBlockDevice_snapshotId :: Lens' EbsBlockDevice (Maybe Text)

The ID of the snapshot.

ebsBlockDevice_volumeType :: Lens' EbsBlockDevice (Maybe VolumeType)

The volume type. For more information, see Amazon EBS volume types in the Amazon EC2 User Guide. If the volume type is io1 or io2, you must specify the IOPS that the volume supports.

ebsBlockDevice_iops :: Lens' EbsBlockDevice (Maybe Int)

The number of I/O operations per second (IOPS). For gp3, io1, and io2 volumes, this represents the number of IOPS that are provisioned for the volume. For gp2 volumes, this represents the baseline performance of the volume and the rate at which the volume accumulates I/O credits for bursting.

The following are the supported values for each volume type:

gp3: 3,000-16,000 IOPS

io1: 100-64,000 IOPS

io2: 100-64,000 IOPS

For io1 and io2 volumes, we guarantee 64,000 IOPS only for Instances built on the Nitro System. Other instance families guarantee performance up to 32,000 IOPS.

This parameter is required for io1 and io2 volumes. The default for gp3 volumes is 3,000 IOPS. This parameter is not supported for gp2, st1, sc1, or standard volumes.

ebsBlockDevice_volumeSize :: Lens' EbsBlockDevice (Maybe Int)

The size of the volume, in GiBs. You must specify either a snapshot ID or a volume size. If you specify a snapshot, the default is the snapshot size. You can specify a volume size that is equal to or larger than the snapshot size.

The following are the supported volumes sizes for each volume type:

gp2 and gp3:1-16,384

io1 and io2: 4-16,384

st1 and sc1: 125-16,384

standard: 1-1,024