Network.AWS.EC2.Types.EnableFastSnapshotRestoreSuccessItem

Description

Documentation

data EnableFastSnapshotRestoreSuccessItem

Describes fast snapshot restores that were successfully enabled.

See: newEnableFastSnapshotRestoreSuccessItem smart constructor.

Constructors

EnableFastSnapshotRestoreSuccessItem'

Fields

disablingTime :: Maybe ISO8601

The time at which fast snapshot restores entered the disabling state.

state :: Maybe FastSnapshotRestoreStateCode

The state of fast snapshot restores.

ownerAlias :: Maybe Text

The Amazon Web Services owner alias that enabled fast snapshot restores on the snapshot. This is intended for future use.

disabledTime :: Maybe ISO8601

The time at which fast snapshot restores entered the disabled state.

enabledTime :: Maybe ISO8601

The time at which fast snapshot restores entered the enabled state.

optimizingTime :: Maybe ISO8601

The time at which fast snapshot restores entered the optimizing state.

ownerId :: Maybe Text

The ID of the Amazon Web Services account that enabled fast snapshot restores on the snapshot.

stateTransitionReason :: Maybe Text

The reason for the state transition. The possible values are as follows:

Client.UserInitiated - The state successfully transitioned to enabling or disabling.

Client.UserInitiated - Lifecycle state transition - The state successfully transitioned to optimizing, enabled, or disabled.

availabilityZone :: Maybe Text

The Availability Zone.

snapshotId :: Maybe Text

The ID of the snapshot.

enablingTime :: Maybe ISO8601

The time at which fast snapshot restores entered the enabling state.

newEnableFastSnapshotRestoreSuccessItem :: EnableFastSnapshotRestoreSuccessItem

Create a value of EnableFastSnapshotRestoreSuccessItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:disablingTime:EnableFastSnapshotRestoreSuccessItem', enableFastSnapshotRestoreSuccessItem_disablingTime - The time at which fast snapshot restores entered the disabling state.

$sel:state:EnableFastSnapshotRestoreSuccessItem', enableFastSnapshotRestoreSuccessItem_state - The state of fast snapshot restores.

$sel:ownerAlias:EnableFastSnapshotRestoreSuccessItem', enableFastSnapshotRestoreSuccessItem_ownerAlias - The Amazon Web Services owner alias that enabled fast snapshot restores on the snapshot. This is intended for future use.

$sel:disabledTime:EnableFastSnapshotRestoreSuccessItem', enableFastSnapshotRestoreSuccessItem_disabledTime - The time at which fast snapshot restores entered the disabled state.

$sel:enabledTime:EnableFastSnapshotRestoreSuccessItem', enableFastSnapshotRestoreSuccessItem_enabledTime - The time at which fast snapshot restores entered the enabled state.

$sel:optimizingTime:EnableFastSnapshotRestoreSuccessItem', enableFastSnapshotRestoreSuccessItem_optimizingTime - The time at which fast snapshot restores entered the optimizing state.

$sel:ownerId:EnableFastSnapshotRestoreSuccessItem', enableFastSnapshotRestoreSuccessItem_ownerId - The ID of the Amazon Web Services account that enabled fast snapshot restores on the snapshot.

$sel:stateTransitionReason:EnableFastSnapshotRestoreSuccessItem', enableFastSnapshotRestoreSuccessItem_stateTransitionReason - The reason for the state transition. The possible values are as follows:

Client.UserInitiated - The state successfully transitioned to enabling or disabling.

Client.UserInitiated - Lifecycle state transition - The state successfully transitioned to optimizing, enabled, or disabled.

$sel:availabilityZone:EnableFastSnapshotRestoreSuccessItem', enableFastSnapshotRestoreSuccessItem_availabilityZone - The Availability Zone.

$sel:snapshotId:EnableFastSnapshotRestoreSuccessItem', enableFastSnapshotRestoreSuccessItem_snapshotId - The ID of the snapshot.

$sel:enablingTime:EnableFastSnapshotRestoreSuccessItem', enableFastSnapshotRestoreSuccessItem_enablingTime - The time at which fast snapshot restores entered the enabling state.

enableFastSnapshotRestoreSuccessItem_disablingTime :: Lens' EnableFastSnapshotRestoreSuccessItem (Maybe UTCTime)

The time at which fast snapshot restores entered the disabling state.

enableFastSnapshotRestoreSuccessItem_state :: Lens' EnableFastSnapshotRestoreSuccessItem (Maybe FastSnapshotRestoreStateCode)

The state of fast snapshot restores.

enableFastSnapshotRestoreSuccessItem_ownerAlias :: Lens' EnableFastSnapshotRestoreSuccessItem (Maybe Text)

The Amazon Web Services owner alias that enabled fast snapshot restores on the snapshot. This is intended for future use.

enableFastSnapshotRestoreSuccessItem_disabledTime :: Lens' EnableFastSnapshotRestoreSuccessItem (Maybe UTCTime)

The time at which fast snapshot restores entered the disabled state.

enableFastSnapshotRestoreSuccessItem_enabledTime :: Lens' EnableFastSnapshotRestoreSuccessItem (Maybe UTCTime)

The time at which fast snapshot restores entered the enabled state.

enableFastSnapshotRestoreSuccessItem_optimizingTime :: Lens' EnableFastSnapshotRestoreSuccessItem (Maybe UTCTime)

The time at which fast snapshot restores entered the optimizing state.

enableFastSnapshotRestoreSuccessItem_ownerId :: Lens' EnableFastSnapshotRestoreSuccessItem (Maybe Text)

The ID of the Amazon Web Services account that enabled fast snapshot restores on the snapshot.

enableFastSnapshotRestoreSuccessItem_stateTransitionReason :: Lens' EnableFastSnapshotRestoreSuccessItem (Maybe Text)

The reason for the state transition. The possible values are as follows:

Client.UserInitiated - The state successfully transitioned to enabling or disabling.

Client.UserInitiated - Lifecycle state transition - The state successfully transitioned to optimizing, enabled, or disabled.

enableFastSnapshotRestoreSuccessItem_availabilityZone :: Lens' EnableFastSnapshotRestoreSuccessItem (Maybe Text)

The Availability Zone.

enableFastSnapshotRestoreSuccessItem_snapshotId :: Lens' EnableFastSnapshotRestoreSuccessItem (Maybe Text)

The ID of the snapshot.

enableFastSnapshotRestoreSuccessItem_enablingTime :: Lens' EnableFastSnapshotRestoreSuccessItem (Maybe UTCTime)

The time at which fast snapshot restores entered the enabling state.