Network.AWS.EC2.Types.EventInformation

Description

Documentation

data EventInformation

Describes an EC2 Fleet or Spot Fleet event.

See: newEventInformation smart constructor.

Constructors

EventInformation'

Fields

instanceId :: Maybe Text

The ID of the instance. This information is available only for instanceChange events.

eventDescription :: Maybe Text

The description of the event.

eventSubType :: Maybe Text

The event.

The following are the error events:

iamFleetRoleInvalid - The EC2 Fleet or Spot Fleet did not have the required permissions either to launch or terminate an instance.

spotFleetRequestConfigurationInvalid - The configuration is not valid. For more information, see the description of the event.

spotInstanceCountLimitExceeded - You've reached the limit on the number of Spot Instances that you can launch.

The following are the fleetRequestChange events:

active - The EC2 Fleet or Spot Fleet request has been validated and Amazon EC2 is attempting to maintain the target number of running Spot Instances.

cancelled - The EC2 Fleet or Spot Fleet request is canceled and has no running Spot Instances. The EC2 Fleet or Spot Fleet will be deleted two days after its instances were terminated.

cancelled_running - The EC2 Fleet or Spot Fleet request is canceled and does not launch additional Spot Instances. Existing Spot Instances continue to run until they are interrupted or terminated.

cancelled_terminating - The EC2 Fleet or Spot Fleet request is canceled and its Spot Instances are terminating.

expired - The EC2 Fleet or Spot Fleet request has expired. A subsequent event indicates that the instances were terminated, if the request was created with TerminateInstancesWithExpiration set.

modify_in_progress - A request to modify the EC2 Fleet or Spot Fleet request was accepted and is in progress.

modify_succeeded - The EC2 Fleet or Spot Fleet request was modified.

submitted - The EC2 Fleet or Spot Fleet request is being evaluated and Amazon EC2 is preparing to launch the target number of Spot Instances.

The following are the instanceChange events:

launched - A request was fulfilled and a new instance was launched.

terminated - An instance was terminated by the user.

The following are the Information events:

launchSpecTemporarilyBlacklisted - The configuration is not valid and several attempts to launch instances have failed. For more information, see the description of the event.

launchSpecUnusable - The price in a launch specification is not valid because it is below the Spot price or the Spot price is above the On-Demand price.

fleetProgressHalted - The price in every launch specification is not valid. A launch specification might become valid if the Spot price changes.

newEventInformation :: EventInformation

Create a value of EventInformation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceId:EventInformation', eventInformation_instanceId - The ID of the instance. This information is available only for instanceChange events.

$sel:eventDescription:EventInformation', eventInformation_eventDescription - The description of the event.

$sel:eventSubType:EventInformation', eventInformation_eventSubType - The event.

The following are the error events:

iamFleetRoleInvalid - The EC2 Fleet or Spot Fleet did not have the required permissions either to launch or terminate an instance.

spotFleetRequestConfigurationInvalid - The configuration is not valid. For more information, see the description of the event.

spotInstanceCountLimitExceeded - You've reached the limit on the number of Spot Instances that you can launch.

The following are the fleetRequestChange events:

active - The EC2 Fleet or Spot Fleet request has been validated and Amazon EC2 is attempting to maintain the target number of running Spot Instances.

cancelled - The EC2 Fleet or Spot Fleet request is canceled and has no running Spot Instances. The EC2 Fleet or Spot Fleet will be deleted two days after its instances were terminated.

cancelled_running - The EC2 Fleet or Spot Fleet request is canceled and does not launch additional Spot Instances. Existing Spot Instances continue to run until they are interrupted or terminated.

cancelled_terminating - The EC2 Fleet or Spot Fleet request is canceled and its Spot Instances are terminating.

expired - The EC2 Fleet or Spot Fleet request has expired. A subsequent event indicates that the instances were terminated, if the request was created with TerminateInstancesWithExpiration set.

modify_in_progress - A request to modify the EC2 Fleet or Spot Fleet request was accepted and is in progress.

modify_succeeded - The EC2 Fleet or Spot Fleet request was modified.

submitted - The EC2 Fleet or Spot Fleet request is being evaluated and Amazon EC2 is preparing to launch the target number of Spot Instances.

The following are the instanceChange events:

launched - A request was fulfilled and a new instance was launched.

terminated - An instance was terminated by the user.

The following are the Information events:

launchSpecTemporarilyBlacklisted - The configuration is not valid and several attempts to launch instances have failed. For more information, see the description of the event.

launchSpecUnusable - The price in a launch specification is not valid because it is below the Spot price or the Spot price is above the On-Demand price.

fleetProgressHalted - The price in every launch specification is not valid. A launch specification might become valid if the Spot price changes.

eventInformation_instanceId :: Lens' EventInformation (Maybe Text)

The ID of the instance. This information is available only for instanceChange events.

eventInformation_eventDescription :: Lens' EventInformation (Maybe Text)

The description of the event.

eventInformation_eventSubType :: Lens' EventInformation (Maybe Text)

The event.

The following are the error events:

iamFleetRoleInvalid - The EC2 Fleet or Spot Fleet did not have the required permissions either to launch or terminate an instance.

spotFleetRequestConfigurationInvalid - The configuration is not valid. For more information, see the description of the event.

spotInstanceCountLimitExceeded - You've reached the limit on the number of Spot Instances that you can launch.

The following are the fleetRequestChange events:

active - The EC2 Fleet or Spot Fleet request has been validated and Amazon EC2 is attempting to maintain the target number of running Spot Instances.

cancelled - The EC2 Fleet or Spot Fleet request is canceled and has no running Spot Instances. The EC2 Fleet or Spot Fleet will be deleted two days after its instances were terminated.

cancelled_running - The EC2 Fleet or Spot Fleet request is canceled and does not launch additional Spot Instances. Existing Spot Instances continue to run until they are interrupted or terminated.

cancelled_terminating - The EC2 Fleet or Spot Fleet request is canceled and its Spot Instances are terminating.

expired - The EC2 Fleet or Spot Fleet request has expired. A subsequent event indicates that the instances were terminated, if the request was created with TerminateInstancesWithExpiration set.

modify_in_progress - A request to modify the EC2 Fleet or Spot Fleet request was accepted and is in progress.

modify_succeeded - The EC2 Fleet or Spot Fleet request was modified.

submitted - The EC2 Fleet or Spot Fleet request is being evaluated and Amazon EC2 is preparing to launch the target number of Spot Instances.

The following are the instanceChange events:

launched - A request was fulfilled and a new instance was launched.

terminated - An instance was terminated by the user.

The following are the Information events:

launchSpecTemporarilyBlacklisted - The configuration is not valid and several attempts to launch instances have failed. For more information, see the description of the event.

launchSpecUnusable - The price in a launch specification is not valid because it is below the Spot price or the Spot price is above the On-Demand price.

fleetProgressHalted - The price in every launch specification is not valid. A launch specification might become valid if the Spot price changes.