Network.AWS.EC2.Types.ExportImageTask

Description

Documentation

data ExportImageTask

Describes an export image task.

See: newExportImageTask smart constructor.

Constructors

ExportImageTask'

Fields

statusMessage :: Maybe Text

The status message for the export image task.

status :: Maybe Text

The status of the export image task. The possible values are active, completed, deleting, and deleted.

imageId :: Maybe Text

The ID of the image.

tags :: Maybe [Tag]

Any tags assigned to the export image task.

s3ExportLocation :: Maybe ExportTaskS3Location

Information about the destination Amazon S3 bucket.

description :: Maybe Text

A description of the image being exported.

exportImageTaskId :: Maybe Text

The ID of the export image task.

progress :: Maybe Text

The percent complete of the export image task.

newExportImageTask :: ExportImageTask

Create a value of ExportImageTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:statusMessage:ExportImageTask', exportImageTask_statusMessage - The status message for the export image task.

$sel:status:ExportImageTask', exportImageTask_status - The status of the export image task. The possible values are active, completed, deleting, and deleted.

$sel:imageId:ExportImageTask', exportImageTask_imageId - The ID of the image.

$sel:tags:ExportImageTask', exportImageTask_tags - Any tags assigned to the export image task.

$sel:s3ExportLocation:ExportImageTask', exportImageTask_s3ExportLocation - Information about the destination Amazon S3 bucket.

$sel:description:ExportImageTask', exportImageTask_description - A description of the image being exported.

$sel:exportImageTaskId:ExportImageTask', exportImageTask_exportImageTaskId - The ID of the export image task.

$sel:progress:ExportImageTask', exportImageTask_progress - The percent complete of the export image task.

exportImageTask_statusMessage :: Lens' ExportImageTask (Maybe Text)

The status message for the export image task.

exportImageTask_status :: Lens' ExportImageTask (Maybe Text)

The status of the export image task. The possible values are active, completed, deleting, and deleted.

exportImageTask_imageId :: Lens' ExportImageTask (Maybe Text)

The ID of the image.

exportImageTask_tags :: Lens' ExportImageTask (Maybe [Tag])

Any tags assigned to the export image task.

exportImageTask_s3ExportLocation :: Lens' ExportImageTask (Maybe ExportTaskS3Location)

Information about the destination Amazon S3 bucket.

exportImageTask_description :: Lens' ExportImageTask (Maybe Text)

A description of the image being exported.

exportImageTask_exportImageTaskId :: Lens' ExportImageTask (Maybe Text)

The ID of the export image task.

exportImageTask_progress :: Lens' ExportImageTask (Maybe Text)

The percent complete of the export image task.