Network.AWS.EC2.Types.FleetCapacityReservation

Description

Documentation

data FleetCapacityReservation

Information about a Capacity Reservation in a Capacity Reservation Fleet.

See: newFleetCapacityReservation smart constructor.

Constructors

FleetCapacityReservation'

Fields

priority :: Maybe Natural

The priority of the instance type in the Capacity Reservation Fleet. For more information, see Instance type priority in the Amazon EC2 User Guide.

availabilityZoneId :: Maybe Text

The ID of the Availability Zone in which the Capacity Reservation reserves capacity.

createDate :: Maybe ISO8601

The date and time at which the Capacity Reservation was created.

weight :: Maybe Double

The weight of the instance type in the Capacity Reservation Fleet. For more information, see Instance type weight in the Amazon EC2 User Guide.

instancePlatform :: Maybe CapacityReservationInstancePlatform

The type of operating system for which the Capacity Reservation reserves capacity.

capacityReservationId :: Maybe Text

The ID of the Capacity Reservation.

instanceType :: Maybe InstanceType

The instance type for which the Capacity Reservation reserves capacity.

ebsOptimized :: Maybe Bool

Indicates whether the Capacity Reservation reserves capacity for EBS-optimized instance types.

availabilityZone :: Maybe Text

The Availability Zone in which the Capacity Reservation reserves capacity.

fulfilledCapacity :: Maybe Double

The number of capacity units fulfilled by the Capacity Reservation. For more information, see Total target capacity in the Amazon EC2 User Guide.

totalInstanceCount :: Maybe Int

The total number of instances for which the Capacity Reservation reserves capacity.

newFleetCapacityReservation :: FleetCapacityReservation

Create a value of FleetCapacityReservation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:priority:FleetCapacityReservation', fleetCapacityReservation_priority - The priority of the instance type in the Capacity Reservation Fleet. For more information, see Instance type priority in the Amazon EC2 User Guide.

$sel:availabilityZoneId:FleetCapacityReservation', fleetCapacityReservation_availabilityZoneId - The ID of the Availability Zone in which the Capacity Reservation reserves capacity.

$sel:createDate:FleetCapacityReservation', fleetCapacityReservation_createDate - The date and time at which the Capacity Reservation was created.

$sel:weight:FleetCapacityReservation', fleetCapacityReservation_weight - The weight of the instance type in the Capacity Reservation Fleet. For more information, see Instance type weight in the Amazon EC2 User Guide.

$sel:instancePlatform:FleetCapacityReservation', fleetCapacityReservation_instancePlatform - The type of operating system for which the Capacity Reservation reserves capacity.

$sel:capacityReservationId:FleetCapacityReservation', fleetCapacityReservation_capacityReservationId - The ID of the Capacity Reservation.

$sel:instanceType:FleetCapacityReservation', fleetCapacityReservation_instanceType - The instance type for which the Capacity Reservation reserves capacity.

$sel:ebsOptimized:FleetCapacityReservation', fleetCapacityReservation_ebsOptimized - Indicates whether the Capacity Reservation reserves capacity for EBS-optimized instance types.

$sel:availabilityZone:FleetCapacityReservation', fleetCapacityReservation_availabilityZone - The Availability Zone in which the Capacity Reservation reserves capacity.

$sel:fulfilledCapacity:FleetCapacityReservation', fleetCapacityReservation_fulfilledCapacity - The number of capacity units fulfilled by the Capacity Reservation. For more information, see Total target capacity in the Amazon EC2 User Guide.

$sel:totalInstanceCount:FleetCapacityReservation', fleetCapacityReservation_totalInstanceCount - The total number of instances for which the Capacity Reservation reserves capacity.

fleetCapacityReservation_priority :: Lens' FleetCapacityReservation (Maybe Natural)

The priority of the instance type in the Capacity Reservation Fleet. For more information, see Instance type priority in the Amazon EC2 User Guide.

fleetCapacityReservation_availabilityZoneId :: Lens' FleetCapacityReservation (Maybe Text)

The ID of the Availability Zone in which the Capacity Reservation reserves capacity.

fleetCapacityReservation_createDate :: Lens' FleetCapacityReservation (Maybe UTCTime)

The date and time at which the Capacity Reservation was created.

fleetCapacityReservation_weight :: Lens' FleetCapacityReservation (Maybe Double)

The weight of the instance type in the Capacity Reservation Fleet. For more information, see Instance type weight in the Amazon EC2 User Guide.

fleetCapacityReservation_instancePlatform :: Lens' FleetCapacityReservation (Maybe CapacityReservationInstancePlatform)

The type of operating system for which the Capacity Reservation reserves capacity.

fleetCapacityReservation_capacityReservationId :: Lens' FleetCapacityReservation (Maybe Text)

The ID of the Capacity Reservation.

fleetCapacityReservation_instanceType :: Lens' FleetCapacityReservation (Maybe InstanceType)

The instance type for which the Capacity Reservation reserves capacity.

fleetCapacityReservation_ebsOptimized :: Lens' FleetCapacityReservation (Maybe Bool)

Indicates whether the Capacity Reservation reserves capacity for EBS-optimized instance types.

fleetCapacityReservation_availabilityZone :: Lens' FleetCapacityReservation (Maybe Text)

The Availability Zone in which the Capacity Reservation reserves capacity.

fleetCapacityReservation_fulfilledCapacity :: Lens' FleetCapacityReservation (Maybe Double)

The number of capacity units fulfilled by the Capacity Reservation. For more information, see Total target capacity in the Amazon EC2 User Guide.

fleetCapacityReservation_totalInstanceCount :: Lens' FleetCapacityReservation (Maybe Int)

The total number of instances for which the Capacity Reservation reserves capacity.