Network.AWS.EC2.Types.FleetData

Description

Documentation

data FleetData

Describes an EC2 Fleet.

See: newFleetData smart constructor.

Constructors

FleetData'

Fields

launchTemplateConfigs :: Maybe [FleetLaunchTemplateConfig]

The launch template and overrides.

fleetState :: Maybe FleetStateCode

The state of the EC2 Fleet.

onDemandOptions :: Maybe OnDemandOptions

The allocation strategy of On-Demand Instances in an EC2 Fleet.

fleetId :: Maybe Text

The ID of the EC2 Fleet.

fulfilledOnDemandCapacity :: Maybe Double

The number of units fulfilled by this request compared to the set target On-Demand capacity.

validFrom :: Maybe ISO8601

The start date and time of the request, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ). The default is to start fulfilling the request immediately.

replaceUnhealthyInstances :: Maybe Bool

Indicates whether EC2 Fleet should replace unhealthy Spot Instances. Supported only for fleets of type maintain. For more information, see EC2 Fleet health checks in the Amazon EC2 User Guide.

instances :: Maybe [DescribeFleetsInstances]

Information about the instances that were launched by the fleet. Valid only when Type is set to instant.

context :: Maybe Text

Reserved.

validUntil :: Maybe ISO8601

The end date and time of the request, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ). At this point, no new instance requests are placed or able to fulfill the request. The default end date is 7 days from the current date.

createTime :: Maybe ISO8601

The creation date and time of the EC2 Fleet.

tags :: Maybe [Tag]

The tags for an EC2 Fleet resource.

activityStatus :: Maybe FleetActivityStatus

The progress of the EC2 Fleet. If there is an error, the status is error. After all requests are placed, the status is pending_fulfillment. If the size of the EC2 Fleet is equal to or greater than its target capacity, the status is fulfilled. If the size of the EC2 Fleet is decreased, the status is pending_termination while instances are terminating.

excessCapacityTerminationPolicy :: Maybe FleetExcessCapacityTerminationPolicy

Indicates whether running instances should be terminated if the target capacity of the EC2 Fleet is decreased below the current size of the EC2 Fleet.

errors :: Maybe [DescribeFleetError]

Information about the instances that could not be launched by the fleet. Valid only when Type is set to instant.

targetCapacitySpecification :: Maybe TargetCapacitySpecification

The number of units to request. You can choose to set the target capacity in terms of instances or a performance characteristic that is important to your application workload, such as vCPUs, memory, or I/O. If the request type is maintain, you can specify a target capacity of 0 and add capacity later.

spotOptions :: Maybe SpotOptions

The configuration of Spot Instances in an EC2 Fleet.

type' :: Maybe FleetType

The type of request. Indicates whether the EC2 Fleet only requests the target capacity, or also attempts to maintain it. If you request a certain target capacity, EC2 Fleet only places the required requests; it does not attempt to replenish instances if capacity is diminished, and it does not submit requests in alternative capacity pools if capacity is unavailable. To maintain a certain target capacity, EC2 Fleet places the required requests to meet this target capacity. It also automatically replenishes any interrupted Spot Instances. Default: maintain.

fulfilledCapacity :: Maybe Double

The number of units fulfilled by this request compared to the set target capacity.

terminateInstancesWithExpiration :: Maybe Bool

Indicates whether running instances should be terminated when the EC2 Fleet expires.

clientToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see Ensuring Idempotency.

Constraints: Maximum 64 ASCII characters

newFleetData :: FleetData

Create a value of FleetData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:launchTemplateConfigs:FleetData', fleetData_launchTemplateConfigs - The launch template and overrides.

$sel:fleetState:FleetData', fleetData_fleetState - The state of the EC2 Fleet.

$sel:onDemandOptions:FleetData', fleetData_onDemandOptions - The allocation strategy of On-Demand Instances in an EC2 Fleet.

$sel:fleetId:FleetData', fleetData_fleetId - The ID of the EC2 Fleet.

$sel:fulfilledOnDemandCapacity:FleetData', fleetData_fulfilledOnDemandCapacity - The number of units fulfilled by this request compared to the set target On-Demand capacity.

$sel:validFrom:FleetData', fleetData_validFrom - The start date and time of the request, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ). The default is to start fulfilling the request immediately.

$sel:replaceUnhealthyInstances:FleetData', fleetData_replaceUnhealthyInstances - Indicates whether EC2 Fleet should replace unhealthy Spot Instances. Supported only for fleets of type maintain. For more information, see EC2 Fleet health checks in the Amazon EC2 User Guide.

$sel:instances:FleetData', fleetData_instances - Information about the instances that were launched by the fleet. Valid only when Type is set to instant.

$sel:context:FleetData', fleetData_context - Reserved.

$sel:validUntil:FleetData', fleetData_validUntil - The end date and time of the request, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ). At this point, no new instance requests are placed or able to fulfill the request. The default end date is 7 days from the current date.

$sel:createTime:FleetData', fleetData_createTime - The creation date and time of the EC2 Fleet.

$sel:tags:FleetData', fleetData_tags - The tags for an EC2 Fleet resource.

$sel:activityStatus:FleetData', fleetData_activityStatus - The progress of the EC2 Fleet. If there is an error, the status is error. After all requests are placed, the status is pending_fulfillment. If the size of the EC2 Fleet is equal to or greater than its target capacity, the status is fulfilled. If the size of the EC2 Fleet is decreased, the status is pending_termination while instances are terminating.

$sel:excessCapacityTerminationPolicy:FleetData', fleetData_excessCapacityTerminationPolicy - Indicates whether running instances should be terminated if the target capacity of the EC2 Fleet is decreased below the current size of the EC2 Fleet.

$sel:errors:FleetData', fleetData_errors - Information about the instances that could not be launched by the fleet. Valid only when Type is set to instant.

$sel:targetCapacitySpecification:FleetData', fleetData_targetCapacitySpecification - The number of units to request. You can choose to set the target capacity in terms of instances or a performance characteristic that is important to your application workload, such as vCPUs, memory, or I/O. If the request type is maintain, you can specify a target capacity of 0 and add capacity later.

$sel:spotOptions:FleetData', fleetData_spotOptions - The configuration of Spot Instances in an EC2 Fleet.

$sel:type':FleetData', fleetData_type - The type of request. Indicates whether the EC2 Fleet only requests the target capacity, or also attempts to maintain it. If you request a certain target capacity, EC2 Fleet only places the required requests; it does not attempt to replenish instances if capacity is diminished, and it does not submit requests in alternative capacity pools if capacity is unavailable. To maintain a certain target capacity, EC2 Fleet places the required requests to meet this target capacity. It also automatically replenishes any interrupted Spot Instances. Default: maintain.

$sel:fulfilledCapacity:FleetData', fleetData_fulfilledCapacity - The number of units fulfilled by this request compared to the set target capacity.

$sel:terminateInstancesWithExpiration:FleetData', fleetData_terminateInstancesWithExpiration - Indicates whether running instances should be terminated when the EC2 Fleet expires.

$sel:clientToken:FleetData', fleetData_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see Ensuring Idempotency.

Constraints: Maximum 64 ASCII characters

fleetData_launchTemplateConfigs :: Lens' FleetData (Maybe [FleetLaunchTemplateConfig])

The launch template and overrides.

fleetData_fleetState :: Lens' FleetData (Maybe FleetStateCode)

The state of the EC2 Fleet.

fleetData_onDemandOptions :: Lens' FleetData (Maybe OnDemandOptions)

The allocation strategy of On-Demand Instances in an EC2 Fleet.

fleetData_fleetId :: Lens' FleetData (Maybe Text)

The ID of the EC2 Fleet.

fleetData_fulfilledOnDemandCapacity :: Lens' FleetData (Maybe Double)

The number of units fulfilled by this request compared to the set target On-Demand capacity.

fleetData_validFrom :: Lens' FleetData (Maybe UTCTime)

The start date and time of the request, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ). The default is to start fulfilling the request immediately.

fleetData_replaceUnhealthyInstances :: Lens' FleetData (Maybe Bool)

Indicates whether EC2 Fleet should replace unhealthy Spot Instances. Supported only for fleets of type maintain. For more information, see EC2 Fleet health checks in the Amazon EC2 User Guide.

fleetData_instances :: Lens' FleetData (Maybe [DescribeFleetsInstances])

Information about the instances that were launched by the fleet. Valid only when Type is set to instant.

fleetData_context :: Lens' FleetData (Maybe Text)

Reserved.

fleetData_validUntil :: Lens' FleetData (Maybe UTCTime)

The end date and time of the request, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ). At this point, no new instance requests are placed or able to fulfill the request. The default end date is 7 days from the current date.

fleetData_createTime :: Lens' FleetData (Maybe UTCTime)

The creation date and time of the EC2 Fleet.

fleetData_tags :: Lens' FleetData (Maybe [Tag])

The tags for an EC2 Fleet resource.

fleetData_activityStatus :: Lens' FleetData (Maybe FleetActivityStatus)

The progress of the EC2 Fleet. If there is an error, the status is error. After all requests are placed, the status is pending_fulfillment. If the size of the EC2 Fleet is equal to or greater than its target capacity, the status is fulfilled. If the size of the EC2 Fleet is decreased, the status is pending_termination while instances are terminating.

fleetData_excessCapacityTerminationPolicy :: Lens' FleetData (Maybe FleetExcessCapacityTerminationPolicy)

Indicates whether running instances should be terminated if the target capacity of the EC2 Fleet is decreased below the current size of the EC2 Fleet.

fleetData_errors :: Lens' FleetData (Maybe [DescribeFleetError])

Information about the instances that could not be launched by the fleet. Valid only when Type is set to instant.

fleetData_targetCapacitySpecification :: Lens' FleetData (Maybe TargetCapacitySpecification)

The number of units to request. You can choose to set the target capacity in terms of instances or a performance characteristic that is important to your application workload, such as vCPUs, memory, or I/O. If the request type is maintain, you can specify a target capacity of 0 and add capacity later.

fleetData_spotOptions :: Lens' FleetData (Maybe SpotOptions)

The configuration of Spot Instances in an EC2 Fleet.

fleetData_type :: Lens' FleetData (Maybe FleetType)

The type of request. Indicates whether the EC2 Fleet only requests the target capacity, or also attempts to maintain it. If you request a certain target capacity, EC2 Fleet only places the required requests; it does not attempt to replenish instances if capacity is diminished, and it does not submit requests in alternative capacity pools if capacity is unavailable. To maintain a certain target capacity, EC2 Fleet places the required requests to meet this target capacity. It also automatically replenishes any interrupted Spot Instances. Default: maintain.

fleetData_fulfilledCapacity :: Lens' FleetData (Maybe Double)

The number of units fulfilled by this request compared to the set target capacity.

fleetData_terminateInstancesWithExpiration :: Lens' FleetData (Maybe Bool)

Indicates whether running instances should be terminated when the EC2 Fleet expires.

fleetData_clientToken :: Lens' FleetData (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see Ensuring Idempotency.

Constraints: Maximum 64 ASCII characters