Network.AWS.EC2.Types.FleetLaunchTemplateOverridesRequest

Description

Documentation

data FleetLaunchTemplateOverridesRequest

Describes overrides for a launch template.

See: newFleetLaunchTemplateOverridesRequest smart constructor.

Constructors

FleetLaunchTemplateOverridesRequest'

Fields

placement :: Maybe Placement

The location where the instance launched, if applicable.

instanceType :: Maybe InstanceType

The instance type.

priority :: Maybe Double

The priority for the launch template override. The highest priority is launched first.

If the On-Demand AllocationStrategy is set to prioritized, EC2 Fleet uses priority to determine which launch template override to use first in fulfilling On-Demand capacity.

If the Spot AllocationStrategy is set to capacity-optimized-prioritized, EC2 Fleet uses priority on a best-effort basis to determine which launch template override to use in fulfilling Spot capacity, but optimizes for capacity first.

Valid values are whole numbers starting at 0. The lower the number, the higher the priority. If no number is set, the launch template override has the lowest priority. You can set the same priority for different launch template overrides.

availabilityZone :: Maybe Text

The Availability Zone in which to launch the instances.

maxPrice :: Maybe Text

The maximum price per unit hour that you are willing to pay for a Spot Instance.

subnetId :: Maybe Text

The IDs of the subnets in which to launch the instances. Separate multiple subnet IDs using commas (for example, subnet-1234abcdeexample1, subnet-0987cdef6example2). A request of type instant can have only one subnet ID.

weightedCapacity :: Maybe Double

The number of units provided by the specified instance type.

newFleetLaunchTemplateOverridesRequest :: FleetLaunchTemplateOverridesRequest

Create a value of FleetLaunchTemplateOverridesRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:placement:FleetLaunchTemplateOverridesRequest', fleetLaunchTemplateOverridesRequest_placement - The location where the instance launched, if applicable.

$sel:instanceType:FleetLaunchTemplateOverridesRequest', fleetLaunchTemplateOverridesRequest_instanceType - The instance type.

$sel:priority:FleetLaunchTemplateOverridesRequest', fleetLaunchTemplateOverridesRequest_priority - The priority for the launch template override. The highest priority is launched first.

If the On-Demand AllocationStrategy is set to prioritized, EC2 Fleet uses priority to determine which launch template override to use first in fulfilling On-Demand capacity.

If the Spot AllocationStrategy is set to capacity-optimized-prioritized, EC2 Fleet uses priority on a best-effort basis to determine which launch template override to use in fulfilling Spot capacity, but optimizes for capacity first.

Valid values are whole numbers starting at 0. The lower the number, the higher the priority. If no number is set, the launch template override has the lowest priority. You can set the same priority for different launch template overrides.

$sel:availabilityZone:FleetLaunchTemplateOverridesRequest', fleetLaunchTemplateOverridesRequest_availabilityZone - The Availability Zone in which to launch the instances.

$sel:maxPrice:FleetLaunchTemplateOverridesRequest', fleetLaunchTemplateOverridesRequest_maxPrice - The maximum price per unit hour that you are willing to pay for a Spot Instance.

$sel:subnetId:FleetLaunchTemplateOverridesRequest', fleetLaunchTemplateOverridesRequest_subnetId - The IDs of the subnets in which to launch the instances. Separate multiple subnet IDs using commas (for example, subnet-1234abcdeexample1, subnet-0987cdef6example2). A request of type instant can have only one subnet ID.

$sel:weightedCapacity:FleetLaunchTemplateOverridesRequest', fleetLaunchTemplateOverridesRequest_weightedCapacity - The number of units provided by the specified instance type.

fleetLaunchTemplateOverridesRequest_placement :: Lens' FleetLaunchTemplateOverridesRequest (Maybe Placement)

The location where the instance launched, if applicable.

fleetLaunchTemplateOverridesRequest_instanceType :: Lens' FleetLaunchTemplateOverridesRequest (Maybe InstanceType)

The instance type.

fleetLaunchTemplateOverridesRequest_priority :: Lens' FleetLaunchTemplateOverridesRequest (Maybe Double)

The priority for the launch template override. The highest priority is launched first.

If the On-Demand AllocationStrategy is set to prioritized, EC2 Fleet uses priority to determine which launch template override to use first in fulfilling On-Demand capacity.

If the Spot AllocationStrategy is set to capacity-optimized-prioritized, EC2 Fleet uses priority on a best-effort basis to determine which launch template override to use in fulfilling Spot capacity, but optimizes for capacity first.

Valid values are whole numbers starting at 0. The lower the number, the higher the priority. If no number is set, the launch template override has the lowest priority. You can set the same priority for different launch template overrides.

fleetLaunchTemplateOverridesRequest_availabilityZone :: Lens' FleetLaunchTemplateOverridesRequest (Maybe Text)

The Availability Zone in which to launch the instances.

fleetLaunchTemplateOverridesRequest_maxPrice :: Lens' FleetLaunchTemplateOverridesRequest (Maybe Text)

The maximum price per unit hour that you are willing to pay for a Spot Instance.

fleetLaunchTemplateOverridesRequest_subnetId :: Lens' FleetLaunchTemplateOverridesRequest (Maybe Text)

The IDs of the subnets in which to launch the instances. Separate multiple subnet IDs using commas (for example, subnet-1234abcdeexample1, subnet-0987cdef6example2). A request of type instant can have only one subnet ID.

fleetLaunchTemplateOverridesRequest_weightedCapacity :: Lens' FleetLaunchTemplateOverridesRequest (Maybe Double)

The number of units provided by the specified instance type.