Network.AWS.EC2.Types.FpgaImage

Description

Documentation

data FpgaImage

Describes an Amazon FPGA image (AFI).

See: newFpgaImage smart constructor.

Constructors

FpgaImage'

Fields

ownerAlias :: Maybe Text

The alias of the AFI owner. Possible values include self, amazon, and aws-marketplace.

ownerId :: Maybe Text

The AWS account ID of the AFI owner.

shellVersion :: Maybe Text

The version of the AWS Shell that was used to create the bitstream.

fpgaImageGlobalId :: Maybe Text

The global FPGA image identifier (AGFI ID).

dataRetentionSupport :: Maybe Bool

Indicates whether data retention support is enabled for the AFI.

productCodes :: Maybe [ProductCode]

The product codes for the AFI.

name :: Maybe Text

The name of the AFI.

pciId :: Maybe PciId

Information about the PCI bus.

state :: Maybe FpgaImageState

Information about the state of the AFI.

updateTime :: Maybe ISO8601

The time of the most recent update to the AFI.

createTime :: Maybe ISO8601

The date and time the AFI was created.

tags :: Maybe [Tag]

Any tags assigned to the AFI.

public :: Maybe Bool

Indicates whether the AFI is public.

description :: Maybe Text

The description of the AFI.

fpgaImageId :: Maybe Text

The FPGA image identifier (AFI ID).

newFpgaImage :: FpgaImage

Create a value of FpgaImage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ownerAlias:FpgaImage', fpgaImage_ownerAlias - The alias of the AFI owner. Possible values include self, amazon, and aws-marketplace.

$sel:ownerId:FpgaImage', fpgaImage_ownerId - The AWS account ID of the AFI owner.

$sel:shellVersion:FpgaImage', fpgaImage_shellVersion - The version of the AWS Shell that was used to create the bitstream.

$sel:fpgaImageGlobalId:FpgaImage', fpgaImage_fpgaImageGlobalId - The global FPGA image identifier (AGFI ID).

$sel:dataRetentionSupport:FpgaImage', fpgaImage_dataRetentionSupport - Indicates whether data retention support is enabled for the AFI.

$sel:productCodes:FpgaImage', fpgaImage_productCodes - The product codes for the AFI.

$sel:name:FpgaImage', fpgaImage_name - The name of the AFI.

$sel:pciId:FpgaImage', fpgaImage_pciId - Information about the PCI bus.

$sel:state:FpgaImage', fpgaImage_state - Information about the state of the AFI.

$sel:updateTime:FpgaImage', fpgaImage_updateTime - The time of the most recent update to the AFI.

$sel:createTime:FpgaImage', fpgaImage_createTime - The date and time the AFI was created.

$sel:tags:FpgaImage', fpgaImage_tags - Any tags assigned to the AFI.

$sel:public:FpgaImage', fpgaImage_public - Indicates whether the AFI is public.

$sel:description:FpgaImage', fpgaImage_description - The description of the AFI.

$sel:fpgaImageId:FpgaImage', fpgaImage_fpgaImageId - The FPGA image identifier (AFI ID).

fpgaImage_ownerAlias :: Lens' FpgaImage (Maybe Text)

The alias of the AFI owner. Possible values include self, amazon, and aws-marketplace.

fpgaImage_ownerId :: Lens' FpgaImage (Maybe Text)

The AWS account ID of the AFI owner.

fpgaImage_shellVersion :: Lens' FpgaImage (Maybe Text)

The version of the AWS Shell that was used to create the bitstream.

fpgaImage_fpgaImageGlobalId :: Lens' FpgaImage (Maybe Text)

The global FPGA image identifier (AGFI ID).

fpgaImage_dataRetentionSupport :: Lens' FpgaImage (Maybe Bool)

Indicates whether data retention support is enabled for the AFI.

fpgaImage_productCodes :: Lens' FpgaImage (Maybe [ProductCode])

The product codes for the AFI.

fpgaImage_name :: Lens' FpgaImage (Maybe Text)

The name of the AFI.

fpgaImage_pciId :: Lens' FpgaImage (Maybe PciId)

Information about the PCI bus.

fpgaImage_state :: Lens' FpgaImage (Maybe FpgaImageState)

Information about the state of the AFI.

fpgaImage_updateTime :: Lens' FpgaImage (Maybe UTCTime)

The time of the most recent update to the AFI.

fpgaImage_createTime :: Lens' FpgaImage (Maybe UTCTime)

The date and time the AFI was created.

fpgaImage_tags :: Lens' FpgaImage (Maybe [Tag])

Any tags assigned to the AFI.

fpgaImage_public :: Lens' FpgaImage (Maybe Bool)

Indicates whether the AFI is public.

fpgaImage_description :: Lens' FpgaImage (Maybe Text)

The description of the AFI.

fpgaImage_fpgaImageId :: Lens' FpgaImage (Maybe Text)

The FPGA image identifier (AFI ID).