Network.AWS.EC2.Types.HostReservation

Description

Documentation

data HostReservation

Details about the Dedicated Host Reservation and associated Dedicated Hosts.

See: newHostReservation smart constructor.

Constructors

HostReservation'

Fields

end :: Maybe ISO8601

The date and time that the reservation ends.

hostIdSet :: Maybe [Text]

The IDs of the Dedicated Hosts associated with the reservation.

instanceFamily :: Maybe Text

The instance family of the Dedicated Host Reservation. The instance family on the Dedicated Host must be the same in order for it to benefit from the reservation.

paymentOption :: Maybe PaymentOption

The payment option selected for this reservation.

upfrontPrice :: Maybe Text

The upfront price of the reservation.

duration :: Maybe Int

The length of the reservation's term, specified in seconds. Can be 31536000 (1 year) | 94608000 (3 years).

hostReservationId :: Maybe Text

The ID of the reservation that specifies the associated Dedicated Hosts.

offeringId :: Maybe Text

The ID of the reservation. This remains the same regardless of which Dedicated Hosts are associated with it.

currencyCode :: Maybe CurrencyCodeValues

The currency in which the upfrontPrice and hourlyPrice amounts are specified. At this time, the only supported currency is USD.

state :: Maybe ReservationState

The state of the reservation.

tags :: Maybe [Tag]

Any tags assigned to the Dedicated Host Reservation.

count :: Maybe Int

The number of Dedicated Hosts the reservation is associated with.

start :: Maybe ISO8601

The date and time that the reservation started.

hourlyPrice :: Maybe Text

The hourly price of the reservation.

newHostReservation :: HostReservation

Create a value of HostReservation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:end:HostReservation', hostReservation_end - The date and time that the reservation ends.

$sel:hostIdSet:HostReservation', hostReservation_hostIdSet - The IDs of the Dedicated Hosts associated with the reservation.

$sel:instanceFamily:HostReservation', hostReservation_instanceFamily - The instance family of the Dedicated Host Reservation. The instance family on the Dedicated Host must be the same in order for it to benefit from the reservation.

$sel:paymentOption:HostReservation', hostReservation_paymentOption - The payment option selected for this reservation.

$sel:upfrontPrice:HostReservation', hostReservation_upfrontPrice - The upfront price of the reservation.

$sel:duration:HostReservation', hostReservation_duration - The length of the reservation's term, specified in seconds. Can be 31536000 (1 year) | 94608000 (3 years).

$sel:hostReservationId:HostReservation', hostReservation_hostReservationId - The ID of the reservation that specifies the associated Dedicated Hosts.

$sel:offeringId:HostReservation', hostReservation_offeringId - The ID of the reservation. This remains the same regardless of which Dedicated Hosts are associated with it.

$sel:currencyCode:HostReservation', hostReservation_currencyCode - The currency in which the upfrontPrice and hourlyPrice amounts are specified. At this time, the only supported currency is USD.

$sel:state:HostReservation', hostReservation_state - The state of the reservation.

$sel:tags:HostReservation', hostReservation_tags - Any tags assigned to the Dedicated Host Reservation.

$sel:count:HostReservation', hostReservation_count - The number of Dedicated Hosts the reservation is associated with.

$sel:start:HostReservation', hostReservation_start - The date and time that the reservation started.

$sel:hourlyPrice:HostReservation', hostReservation_hourlyPrice - The hourly price of the reservation.

hostReservation_end :: Lens' HostReservation (Maybe UTCTime)

The date and time that the reservation ends.

hostReservation_hostIdSet :: Lens' HostReservation (Maybe [Text])

The IDs of the Dedicated Hosts associated with the reservation.

hostReservation_instanceFamily :: Lens' HostReservation (Maybe Text)

The instance family of the Dedicated Host Reservation. The instance family on the Dedicated Host must be the same in order for it to benefit from the reservation.

hostReservation_paymentOption :: Lens' HostReservation (Maybe PaymentOption)

The payment option selected for this reservation.

hostReservation_upfrontPrice :: Lens' HostReservation (Maybe Text)

The upfront price of the reservation.

hostReservation_duration :: Lens' HostReservation (Maybe Int)

The length of the reservation's term, specified in seconds. Can be 31536000 (1 year) | 94608000 (3 years).

hostReservation_hostReservationId :: Lens' HostReservation (Maybe Text)

The ID of the reservation that specifies the associated Dedicated Hosts.

hostReservation_offeringId :: Lens' HostReservation (Maybe Text)

The ID of the reservation. This remains the same regardless of which Dedicated Hosts are associated with it.

hostReservation_currencyCode :: Lens' HostReservation (Maybe CurrencyCodeValues)

The currency in which the upfrontPrice and hourlyPrice amounts are specified. At this time, the only supported currency is USD.

hostReservation_state :: Lens' HostReservation (Maybe ReservationState)

The state of the reservation.

hostReservation_tags :: Lens' HostReservation (Maybe [Tag])

Any tags assigned to the Dedicated Host Reservation.

hostReservation_count :: Lens' HostReservation (Maybe Int)

The number of Dedicated Hosts the reservation is associated with.

hostReservation_start :: Lens' HostReservation (Maybe UTCTime)

The date and time that the reservation started.

hostReservation_hourlyPrice :: Lens' HostReservation (Maybe Text)

The hourly price of the reservation.