Network.AWS.EC2.Types.Image

Description

Documentation

data Image

Describes an image.

See: newImage smart constructor.

Constructors

Image'

Fields

platform :: Maybe PlatformValues

This value is set to windows for Windows AMIs; otherwise, it is blank.

rootDeviceName :: Maybe Text

The device name of the root device volume (for example, /dev/sda1).

ramdiskId :: Maybe Text

The RAM disk associated with the image, if any. Only applicable for machine images.

stateReason :: Maybe StateReason

The reason for the state change.

usageOperation :: Maybe Text

The operation of the Amazon EC2 instance and the billing code that is associated with the AMI. usageOperation corresponds to the lineitem/Operation column on your Amazon Web Services Cost and Usage Report and in the Amazon Web Services Price List API. You can view these fields on the Instances or AMIs pages in the Amazon EC2 console, or in the responses that are returned by the DescribeImages command in the Amazon EC2 API, or the describe-images command in the CLI.

platformDetails :: Maybe Text

The platform details associated with the billing code of the AMI. For more information, see Understanding AMI billing in the Amazon Elastic Compute Cloud User Guide.

bootMode :: Maybe BootModeValues

The boot mode of the image. For more information, see Boot modes in the Amazon Elastic Compute Cloud User Guide.

creationDate :: Maybe Text

The date and time the image was created.

productCodes :: Maybe [ProductCode]

Any product codes associated with the AMI.

deprecationTime :: Maybe Text

The date and time to deprecate the AMI, in UTC, in the following format: YYYY-MM-DDTHH:MM:SSZ. If you specified a value for seconds, Amazon EC2 rounds the seconds to the nearest minute.

name :: Maybe Text

The name of the AMI that was provided during image creation.

tags :: Maybe [Tag]

Any tags assigned to the image.

sriovNetSupport :: Maybe Text

Specifies whether enhanced networking with the Intel 82599 Virtual Function interface is enabled.

blockDeviceMappings :: Maybe [BlockDeviceMapping]

Any block device mapping entries.

description :: Maybe Text

The description of the AMI that was provided during image creation.

kernelId :: Maybe Text

The kernel associated with the image, if any. Only applicable for machine images.

imageOwnerAlias :: Maybe Text

The Amazon Web Services account alias (for example, amazon, self) or the Amazon Web Services account ID of the AMI owner.

enaSupport :: Maybe Bool

Specifies whether enhanced networking with ENA is enabled.

imageId :: Text

The ID of the AMI.

imageLocation :: Text

The location of the AMI.

state :: ImageState

The current state of the AMI. If the state is available, the image is successfully registered and can be used to launch an instance.

ownerId :: Text

The ID of the Amazon Web Services account that owns the image.

public :: Bool

Indicates whether the image has public launch permissions. The value is true if this image has public launch permissions or false if it has only implicit and explicit launch permissions.

architecture :: ArchitectureValues

The architecture of the image.

imageType :: ImageTypeValues

The type of image.

rootDeviceType :: DeviceType

The type of root device used by the AMI. The AMI can use an Amazon EBS volume or an instance store volume.

virtualizationType :: VirtualizationType

The type of virtualization of the AMI.

hypervisor :: HypervisorType

The hypervisor type of the image.

newImage

Arguments

:: Text
$sel:imageId:Image'

-> Text
$sel:imageLocation:Image'

-> ImageState
$sel:state:Image'

-> Text
$sel:ownerId:Image'

-> Bool
$sel:public:Image'

-> ArchitectureValues
$sel:architecture:Image'

-> ImageTypeValues
$sel:imageType:Image'

-> DeviceType
$sel:rootDeviceType:Image'

-> VirtualizationType
$sel:virtualizationType:Image'

-> HypervisorType
$sel:hypervisor:Image'

-> Image

Create a value of Image with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:platform:Image', image_platform - This value is set to windows for Windows AMIs; otherwise, it is blank.

$sel:rootDeviceName:Image', image_rootDeviceName - The device name of the root device volume (for example, /dev/sda1).

$sel:ramdiskId:Image', image_ramdiskId - The RAM disk associated with the image, if any. Only applicable for machine images.

$sel:stateReason:Image', image_stateReason - The reason for the state change.

$sel:usageOperation:Image', image_usageOperation - The operation of the Amazon EC2 instance and the billing code that is associated with the AMI. usageOperation corresponds to the lineitem/Operation column on your Amazon Web Services Cost and Usage Report and in the Amazon Web Services Price List API. You can view these fields on the Instances or AMIs pages in the Amazon EC2 console, or in the responses that are returned by the DescribeImages command in the Amazon EC2 API, or the describe-images command in the CLI.

$sel:platformDetails:Image', image_platformDetails - The platform details associated with the billing code of the AMI. For more information, see Understanding AMI billing in the Amazon Elastic Compute Cloud User Guide.

$sel:bootMode:Image', image_bootMode - The boot mode of the image. For more information, see Boot modes in the Amazon Elastic Compute Cloud User Guide.

$sel:creationDate:Image', image_creationDate - The date and time the image was created.

$sel:productCodes:Image', image_productCodes - Any product codes associated with the AMI.

$sel:deprecationTime:Image', image_deprecationTime - The date and time to deprecate the AMI, in UTC, in the following format: YYYY-MM-DDTHH:MM:SSZ. If you specified a value for seconds, Amazon EC2 rounds the seconds to the nearest minute.

$sel:name:Image', image_name - The name of the AMI that was provided during image creation.

$sel:tags:Image', image_tags - Any tags assigned to the image.

$sel:sriovNetSupport:Image', image_sriovNetSupport - Specifies whether enhanced networking with the Intel 82599 Virtual Function interface is enabled.

$sel:blockDeviceMappings:Image', image_blockDeviceMappings - Any block device mapping entries.

$sel:description:Image', image_description - The description of the AMI that was provided during image creation.

$sel:kernelId:Image', image_kernelId - The kernel associated with the image, if any. Only applicable for machine images.

$sel:imageOwnerAlias:Image', image_imageOwnerAlias - The Amazon Web Services account alias (for example, amazon, self) or the Amazon Web Services account ID of the AMI owner.

$sel:enaSupport:Image', image_enaSupport - Specifies whether enhanced networking with ENA is enabled.

$sel:imageId:Image', image_imageId - The ID of the AMI.

$sel:imageLocation:Image', image_imageLocation - The location of the AMI.

$sel:state:Image', image_state - The current state of the AMI. If the state is available, the image is successfully registered and can be used to launch an instance.

$sel:ownerId:Image', image_ownerId - The ID of the Amazon Web Services account that owns the image.

$sel:public:Image', image_public - Indicates whether the image has public launch permissions. The value is true if this image has public launch permissions or false if it has only implicit and explicit launch permissions.

$sel:architecture:Image', image_architecture - The architecture of the image.

$sel:imageType:Image', image_imageType - The type of image.

$sel:rootDeviceType:Image', image_rootDeviceType - The type of root device used by the AMI. The AMI can use an Amazon EBS volume or an instance store volume.

$sel:virtualizationType:Image', image_virtualizationType - The type of virtualization of the AMI.

$sel:hypervisor:Image', image_hypervisor - The hypervisor type of the image.

image_platform :: Lens' Image (Maybe PlatformValues)

This value is set to windows for Windows AMIs; otherwise, it is blank.

image_rootDeviceName :: Lens' Image (Maybe Text)

The device name of the root device volume (for example, /dev/sda1).

image_ramdiskId :: Lens' Image (Maybe Text)

The RAM disk associated with the image, if any. Only applicable for machine images.

image_stateReason :: Lens' Image (Maybe StateReason)

The reason for the state change.

image_usageOperation :: Lens' Image (Maybe Text)

The operation of the Amazon EC2 instance and the billing code that is associated with the AMI. usageOperation corresponds to the lineitem/Operation column on your Amazon Web Services Cost and Usage Report and in the Amazon Web Services Price List API. You can view these fields on the Instances or AMIs pages in the Amazon EC2 console, or in the responses that are returned by the DescribeImages command in the Amazon EC2 API, or the describe-images command in the CLI.

image_platformDetails :: Lens' Image (Maybe Text)

The platform details associated with the billing code of the AMI. For more information, see Understanding AMI billing in the Amazon Elastic Compute Cloud User Guide.

image_bootMode :: Lens' Image (Maybe BootModeValues)

The boot mode of the image. For more information, see Boot modes in the Amazon Elastic Compute Cloud User Guide.

image_creationDate :: Lens' Image (Maybe Text)

The date and time the image was created.

image_productCodes :: Lens' Image (Maybe [ProductCode])

Any product codes associated with the AMI.

image_deprecationTime :: Lens' Image (Maybe Text)

The date and time to deprecate the AMI, in UTC, in the following format: YYYY-MM-DDTHH:MM:SSZ. If you specified a value for seconds, Amazon EC2 rounds the seconds to the nearest minute.

image_name :: Lens' Image (Maybe Text)

The name of the AMI that was provided during image creation.

image_tags :: Lens' Image (Maybe [Tag])

Any tags assigned to the image.

image_sriovNetSupport :: Lens' Image (Maybe Text)

Specifies whether enhanced networking with the Intel 82599 Virtual Function interface is enabled.

image_blockDeviceMappings :: Lens' Image (Maybe [BlockDeviceMapping])

Any block device mapping entries.

image_description :: Lens' Image (Maybe Text)

The description of the AMI that was provided during image creation.

image_kernelId :: Lens' Image (Maybe Text)

The kernel associated with the image, if any. Only applicable for machine images.

image_imageOwnerAlias :: Lens' Image (Maybe Text)

The Amazon Web Services account alias (for example, amazon, self) or the Amazon Web Services account ID of the AMI owner.

image_enaSupport :: Lens' Image (Maybe Bool)

Specifies whether enhanced networking with ENA is enabled.

image_imageId :: Lens' Image Text

The ID of the AMI.

image_imageLocation :: Lens' Image Text

The location of the AMI.

image_state :: Lens' Image ImageState

The current state of the AMI. If the state is available, the image is successfully registered and can be used to launch an instance.

image_ownerId :: Lens' Image Text

The ID of the Amazon Web Services account that owns the image.

image_public :: Lens' Image Bool

Indicates whether the image has public launch permissions. The value is true if this image has public launch permissions or false if it has only implicit and explicit launch permissions.

image_architecture :: Lens' Image ArchitectureValues

The architecture of the image.

image_imageType :: Lens' Image ImageTypeValues

The type of image.

image_rootDeviceType :: Lens' Image DeviceType

The type of root device used by the AMI. The AMI can use an Amazon EBS volume or an instance store volume.

image_virtualizationType :: Lens' Image VirtualizationType

The type of virtualization of the AMI.

image_hypervisor :: Lens' Image HypervisorType

The hypervisor type of the image.