Network.AWS.EC2.Types.ImportImageTask

Description

Documentation

data ImportImageTask

Describes an import image task.

See: newImportImageTask smart constructor.

Constructors

ImportImageTask'

Fields

platform :: Maybe Text

The description string for the import image task.

hypervisor :: Maybe Text

The target hypervisor for the import task.

Valid values: xen

statusMessage :: Maybe Text

A descriptive status message for the import image task.

status :: Maybe Text

A brief status for the import image task.

encrypted :: Maybe Bool

Indicates whether the image is encrypted.

snapshotDetails :: Maybe [SnapshotDetail]

Information about the snapshots.

usageOperation :: Maybe Text

The usage operation value.

architecture :: Maybe Text

The architecture of the virtual machine.

Valid values: i386 | x86_64 | arm64

bootMode :: Maybe BootModeValues

The boot mode of the virtual machine.

licenseSpecifications :: Maybe [ImportImageLicenseConfigurationResponse]

The ARNs of the license configurations that are associated with the import image task.

importTaskId :: Maybe Text

The ID of the import image task.

imageId :: Maybe Text

The ID of the Amazon Machine Image (AMI) of the imported virtual machine.

kmsKeyId :: Maybe Text

The identifier for the KMS key that was used to create the encrypted image.

tags :: Maybe [Tag]

The tags for the import image task.

description :: Maybe Text

A description of the import task.

licenseType :: Maybe Text

The license type of the virtual machine.

progress :: Maybe Text

The percentage of progress of the import image task.

newImportImageTask :: ImportImageTask

Create a value of ImportImageTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:platform:ImportImageTask', importImageTask_platform - The description string for the import image task.

$sel:hypervisor:ImportImageTask', importImageTask_hypervisor - The target hypervisor for the import task.

Valid values: xen

$sel:statusMessage:ImportImageTask', importImageTask_statusMessage - A descriptive status message for the import image task.

$sel:status:ImportImageTask', importImageTask_status - A brief status for the import image task.

$sel:encrypted:ImportImageTask', importImageTask_encrypted - Indicates whether the image is encrypted.

$sel:snapshotDetails:ImportImageTask', importImageTask_snapshotDetails - Information about the snapshots.

$sel:usageOperation:ImportImageTask', importImageTask_usageOperation - The usage operation value.

$sel:architecture:ImportImageTask', importImageTask_architecture - The architecture of the virtual machine.

Valid values: i386 | x86_64 | arm64

$sel:bootMode:ImportImageTask', importImageTask_bootMode - The boot mode of the virtual machine.

$sel:licenseSpecifications:ImportImageTask', importImageTask_licenseSpecifications - The ARNs of the license configurations that are associated with the import image task.

$sel:importTaskId:ImportImageTask', importImageTask_importTaskId - The ID of the import image task.

$sel:imageId:ImportImageTask', importImageTask_imageId - The ID of the Amazon Machine Image (AMI) of the imported virtual machine.

$sel:kmsKeyId:ImportImageTask', importImageTask_kmsKeyId - The identifier for the KMS key that was used to create the encrypted image.

$sel:tags:ImportImageTask', importImageTask_tags - The tags for the import image task.

$sel:description:ImportImageTask', importImageTask_description - A description of the import task.

$sel:licenseType:ImportImageTask', importImageTask_licenseType - The license type of the virtual machine.

$sel:progress:ImportImageTask', importImageTask_progress - The percentage of progress of the import image task.

importImageTask_platform :: Lens' ImportImageTask (Maybe Text)

The description string for the import image task.

importImageTask_hypervisor :: Lens' ImportImageTask (Maybe Text)

The target hypervisor for the import task.

Valid values: xen

importImageTask_statusMessage :: Lens' ImportImageTask (Maybe Text)

A descriptive status message for the import image task.

importImageTask_status :: Lens' ImportImageTask (Maybe Text)

A brief status for the import image task.

importImageTask_encrypted :: Lens' ImportImageTask (Maybe Bool)

Indicates whether the image is encrypted.

importImageTask_snapshotDetails :: Lens' ImportImageTask (Maybe [SnapshotDetail])

Information about the snapshots.

importImageTask_usageOperation :: Lens' ImportImageTask (Maybe Text)

The usage operation value.

importImageTask_architecture :: Lens' ImportImageTask (Maybe Text)

The architecture of the virtual machine.

Valid values: i386 | x86_64 | arm64

importImageTask_bootMode :: Lens' ImportImageTask (Maybe BootModeValues)

The boot mode of the virtual machine.

importImageTask_licenseSpecifications :: Lens' ImportImageTask (Maybe [ImportImageLicenseConfigurationResponse])

The ARNs of the license configurations that are associated with the import image task.

importImageTask_importTaskId :: Lens' ImportImageTask (Maybe Text)

The ID of the import image task.

importImageTask_imageId :: Lens' ImportImageTask (Maybe Text)

The ID of the Amazon Machine Image (AMI) of the imported virtual machine.

importImageTask_kmsKeyId :: Lens' ImportImageTask (Maybe Text)

The identifier for the KMS key that was used to create the encrypted image.

importImageTask_tags :: Lens' ImportImageTask (Maybe [Tag])

The tags for the import image task.

importImageTask_description :: Lens' ImportImageTask (Maybe Text)

A description of the import task.

importImageTask_licenseType :: Lens' ImportImageTask (Maybe Text)

The license type of the virtual machine.

importImageTask_progress :: Lens' ImportImageTask (Maybe Text)

The percentage of progress of the import image task.