Network.AWS.EC2.Types.ImportInstanceLaunchSpecification

Description

Documentation

data ImportInstanceLaunchSpecification

Describes the launch specification for VM import.

See: newImportInstanceLaunchSpecification smart constructor.

Constructors

ImportInstanceLaunchSpecification'

Fields

additionalInfo :: Maybe Text

Reserved.

groupNames :: Maybe [Text]

The security group names.

subnetId :: Maybe Text

EC2-VPC

The ID of the subnet in which to launch the instance.

instanceType :: Maybe InstanceType

The instance type. For more information about the instance types that you can import, see Instance Types in the VM Import/Export User Guide.

groupIds :: Maybe [Text]

The security group IDs.

userData :: Maybe (Sensitive UserData)

The Base64-encoded user data to make available to the instance.

monitoring :: Maybe Bool

Indicates whether monitoring is enabled.

privateIpAddress :: Maybe Text

EC2-VPC

An available IP address from the IP address range of the subnet.

instanceInitiatedShutdownBehavior :: Maybe ShutdownBehavior

Indicates whether an instance stops or terminates when you initiate shutdown from the instance (using the operating system command for system shutdown).

architecture :: Maybe ArchitectureValues

The architecture of the instance.

placement :: Maybe Placement

The placement information for the instance.

newImportInstanceLaunchSpecification :: ImportInstanceLaunchSpecification

Create a value of ImportInstanceLaunchSpecification with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:additionalInfo:ImportInstanceLaunchSpecification', importInstanceLaunchSpecification_additionalInfo - Reserved.

$sel:groupNames:ImportInstanceLaunchSpecification', importInstanceLaunchSpecification_groupNames - The security group names.

$sel:subnetId:ImportInstanceLaunchSpecification', importInstanceLaunchSpecification_subnetId - [EC2-VPC] The ID of the subnet in which to launch the instance.

$sel:instanceType:ImportInstanceLaunchSpecification', importInstanceLaunchSpecification_instanceType - The instance type. For more information about the instance types that you can import, see Instance Types in the VM Import/Export User Guide.

$sel:groupIds:ImportInstanceLaunchSpecification', importInstanceLaunchSpecification_groupIds - The security group IDs.

$sel:userData:ImportInstanceLaunchSpecification', importInstanceLaunchSpecification_userData - The Base64-encoded user data to make available to the instance.

$sel:monitoring:ImportInstanceLaunchSpecification', importInstanceLaunchSpecification_monitoring - Indicates whether monitoring is enabled.

$sel:privateIpAddress:ImportInstanceLaunchSpecification', importInstanceLaunchSpecification_privateIpAddress - [EC2-VPC] An available IP address from the IP address range of the subnet.

$sel:instanceInitiatedShutdownBehavior:ImportInstanceLaunchSpecification', importInstanceLaunchSpecification_instanceInitiatedShutdownBehavior - Indicates whether an instance stops or terminates when you initiate shutdown from the instance (using the operating system command for system shutdown).

$sel:architecture:ImportInstanceLaunchSpecification', importInstanceLaunchSpecification_architecture - The architecture of the instance.

$sel:placement:ImportInstanceLaunchSpecification', importInstanceLaunchSpecification_placement - The placement information for the instance.

importInstanceLaunchSpecification_additionalInfo :: Lens' ImportInstanceLaunchSpecification (Maybe Text)

Reserved.

importInstanceLaunchSpecification_groupNames :: Lens' ImportInstanceLaunchSpecification (Maybe [Text])

The security group names.

importInstanceLaunchSpecification_subnetId :: Lens' ImportInstanceLaunchSpecification (Maybe Text)

EC2-VPC

The ID of the subnet in which to launch the instance.

importInstanceLaunchSpecification_instanceType :: Lens' ImportInstanceLaunchSpecification (Maybe InstanceType)

The instance type. For more information about the instance types that you can import, see Instance Types in the VM Import/Export User Guide.

importInstanceLaunchSpecification_groupIds :: Lens' ImportInstanceLaunchSpecification (Maybe [Text])

The security group IDs.

importInstanceLaunchSpecification_userData :: Lens' ImportInstanceLaunchSpecification (Maybe UserData)

The Base64-encoded user data to make available to the instance.

importInstanceLaunchSpecification_monitoring :: Lens' ImportInstanceLaunchSpecification (Maybe Bool)

Indicates whether monitoring is enabled.

importInstanceLaunchSpecification_privateIpAddress :: Lens' ImportInstanceLaunchSpecification (Maybe Text)

EC2-VPC

An available IP address from the IP address range of the subnet.

importInstanceLaunchSpecification_instanceInitiatedShutdownBehavior :: Lens' ImportInstanceLaunchSpecification (Maybe ShutdownBehavior)

Indicates whether an instance stops or terminates when you initiate shutdown from the instance (using the operating system command for system shutdown).

importInstanceLaunchSpecification_architecture :: Lens' ImportInstanceLaunchSpecification (Maybe ArchitectureValues)

The architecture of the instance.

importInstanceLaunchSpecification_placement :: Lens' ImportInstanceLaunchSpecification (Maybe Placement)

The placement information for the instance.