Network.AWS.EC2.Types.Instance

Description

Documentation

data Instance

Describes an instance.

See: newInstance smart constructor.

Constructors

Instance'

Fields

platform :: Maybe PlatformValues

The value is Windows for Windows instances; otherwise blank.

instanceLifecycle :: Maybe InstanceLifecycleType

Indicates whether this is a Spot Instance or a Scheduled Instance.

stateTransitionReason :: Maybe Text

The reason for the most recent state transition. This might be an empty string.

capacityReservationSpecification :: Maybe CapacityReservationSpecificationResponse

Information about the Capacity Reservation targeting option.

ebsOptimized :: Maybe Bool

Indicates whether the instance is optimized for Amazon EBS I/O. This optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal I/O performance. This optimization isn't available with all instance types. Additional usage charges apply when using an EBS Optimized instance.

rootDeviceName :: Maybe Text

The device name of the root device volume (for example, /dev/sda1).

ramdiskId :: Maybe Text

The RAM disk associated with this instance, if applicable.

stateReason :: Maybe StateReason

The reason for the most recent state transition.

outpostArn :: Maybe Text

The Amazon Resource Name (ARN) of the Outpost.

elasticInferenceAcceleratorAssociations :: Maybe [ElasticInferenceAcceleratorAssociation]

The elastic inference accelerator associated with the instance.

usageOperation :: Maybe Text

The usage operation value for the instance. For more information, see AMI billing information fields in the Amazon EC2 User Guide.

platformDetails :: Maybe Text

The platform details value for the instance. For more information, see AMI billing information fields in the Amazon EC2 User Guide.

bootMode :: Maybe BootModeValues

The boot mode of the instance. For more information, see Boot modes in the Amazon EC2 User Guide.

productCodes :: Maybe [ProductCode]

The product codes attached to this instance, if applicable.

sourceDestCheck :: Maybe Bool

Indicates whether source/destination checking is enabled.

securityGroups :: Maybe [GroupIdentifier]

The security groups for the instance.

publicDnsName :: Maybe Text

(IPv4 only) The public DNS name assigned to the instance. This name is not available until the instance enters the running state. For EC2-VPC, this name is only available if you've enabled DNS hostnames for your VPC.

iamInstanceProfile :: Maybe IamInstanceProfile

The IAM instance profile associated with the instance, if applicable.

hibernationOptions :: Maybe HibernationOptions

Indicates whether the instance is enabled for hibernation.

usageOperationUpdateTime :: Maybe ISO8601

The time that the usage operation was last updated.

tags :: Maybe [Tag]

Any tags assigned to the instance.

capacityReservationId :: Maybe Text

The ID of the Capacity Reservation.

sriovNetSupport :: Maybe Text

Specifies whether enhanced networking with the Intel 82599 Virtual Function interface is enabled.

publicIpAddress :: Maybe Text

The public IPv4 address, or the Carrier IP address assigned to the instance, if applicable.

A Carrier IP address only applies to an instance launched in a subnet associated with a Wavelength Zone.

blockDeviceMappings :: Maybe [InstanceBlockDeviceMapping]

Any block device mapping entries for the instance.

subnetId :: Maybe Text

EC2-VPC

The ID of the subnet in which the instance is running.

enclaveOptions :: Maybe EnclaveOptions

Indicates whether the instance is enabled for Amazon Web Services Nitro Enclaves.

kernelId :: Maybe Text

The kernel associated with this instance, if applicable.

cpuOptions :: Maybe CpuOptions

The CPU options for the instance.

privateDnsName :: Maybe Text

(IPv4 only) The private DNS hostname name assigned to the instance. This DNS hostname can only be used inside the Amazon EC2 network. This name is not available until the instance enters the running state.

EC2-VPC

The Amazon-provided DNS server resolves Amazon-provided private DNS hostnames if you've enabled DNS resolution and DNS hostnames in your VPC. If you are not using the Amazon-provided DNS server in your VPC, your custom domain name servers must resolve the hostname as appropriate.

licenses :: Maybe [LicenseConfiguration]

The license configurations for the instance.

vpcId :: Maybe Text

EC2-VPC

The ID of the VPC in which the instance is running.

networkInterfaces :: Maybe [InstanceNetworkInterface]

EC2-VPC

The network interfaces for the instance.

keyName :: Maybe Text

The name of the key pair, if this instance was launched with an associated key pair.

spotInstanceRequestId :: Maybe Text

If the request is a Spot Instance request, the ID of the request.

elasticGpuAssociations :: Maybe [ElasticGpuAssociation]

The Elastic GPU associated with the instance.

enaSupport :: Maybe Bool

Specifies whether enhanced networking with ENA is enabled.

metadataOptions :: Maybe InstanceMetadataOptionsResponse

The metadata options for the instance.

privateIpAddress :: Maybe Text

The private IPv4 address assigned to the instance.

clientToken :: Maybe Text

The idempotency token you provided when you launched the instance, if applicable.

instanceId :: Text

The ID of the instance.

imageId :: Text

The ID of the AMI used to launch the instance.

amiLaunchIndex :: Int

The AMI launch index, which can be used to find this instance in the launch group.

instanceType :: InstanceType

The instance type.

launchTime :: ISO8601

The time the instance was launched.

placement :: Placement

The location where the instance launched, if applicable.

monitoring :: Monitoring

The monitoring for the instance.

architecture :: ArchitectureValues

The architecture of the image.

rootDeviceType :: DeviceType

The root device type used by the AMI. The AMI can use an EBS volume or an instance store volume.

virtualizationType :: VirtualizationType

The virtualization type of the instance.

hypervisor :: HypervisorType

The hypervisor type of the instance. The value xen is used for both Xen and Nitro hypervisors.

state :: InstanceState

The current state of the instance.

newInstance

Arguments

:: Text
$sel:instanceId:Instance'

-> Text
$sel:imageId:Instance'

-> Int
$sel:amiLaunchIndex:Instance'

-> InstanceType
$sel:instanceType:Instance'

-> UTCTime
$sel:launchTime:Instance'

-> Placement
$sel:placement:Instance'

-> Monitoring
$sel:monitoring:Instance'

-> ArchitectureValues
$sel:architecture:Instance'

-> DeviceType
$sel:rootDeviceType:Instance'

-> VirtualizationType
$sel:virtualizationType:Instance'

-> HypervisorType
$sel:hypervisor:Instance'

-> InstanceState
$sel:state:Instance'

-> Instance

Create a value of Instance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:platform:Instance', instance_platform - The value is Windows for Windows instances; otherwise blank.

$sel:instanceLifecycle:Instance', instance_instanceLifecycle - Indicates whether this is a Spot Instance or a Scheduled Instance.

$sel:stateTransitionReason:Instance', instance_stateTransitionReason - The reason for the most recent state transition. This might be an empty string.

$sel:capacityReservationSpecification:Instance', instance_capacityReservationSpecification - Information about the Capacity Reservation targeting option.

$sel:ebsOptimized:Instance', instance_ebsOptimized - Indicates whether the instance is optimized for Amazon EBS I/O. This optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal I/O performance. This optimization isn't available with all instance types. Additional usage charges apply when using an EBS Optimized instance.

$sel:rootDeviceName:Instance', instance_rootDeviceName - The device name of the root device volume (for example, /dev/sda1).

$sel:ramdiskId:Instance', instance_ramdiskId - The RAM disk associated with this instance, if applicable.

$sel:stateReason:Instance', instance_stateReason - The reason for the most recent state transition.

$sel:outpostArn:Instance', instance_outpostArn - The Amazon Resource Name (ARN) of the Outpost.

$sel:elasticInferenceAcceleratorAssociations:Instance', instance_elasticInferenceAcceleratorAssociations - The elastic inference accelerator associated with the instance.

$sel:usageOperation:Instance', instance_usageOperation - The usage operation value for the instance. For more information, see AMI billing information fields in the Amazon EC2 User Guide.

$sel:platformDetails:Instance', instance_platformDetails - The platform details value for the instance. For more information, see AMI billing information fields in the Amazon EC2 User Guide.

$sel:bootMode:Instance', instance_bootMode - The boot mode of the instance. For more information, see Boot modes in the Amazon EC2 User Guide.

$sel:productCodes:Instance', instance_productCodes - The product codes attached to this instance, if applicable.

$sel:sourceDestCheck:Instance', instance_sourceDestCheck - Indicates whether source/destination checking is enabled.

$sel:securityGroups:Instance', instance_securityGroups - The security groups for the instance.

$sel:publicDnsName:Instance', instance_publicDnsName - (IPv4 only) The public DNS name assigned to the instance. This name is not available until the instance enters the running state. For EC2-VPC, this name is only available if you've enabled DNS hostnames for your VPC.

$sel:iamInstanceProfile:Instance', instance_iamInstanceProfile - The IAM instance profile associated with the instance, if applicable.

$sel:hibernationOptions:Instance', instance_hibernationOptions - Indicates whether the instance is enabled for hibernation.

$sel:usageOperationUpdateTime:Instance', instance_usageOperationUpdateTime - The time that the usage operation was last updated.

$sel:tags:Instance', instance_tags - Any tags assigned to the instance.

$sel:capacityReservationId:Instance', instance_capacityReservationId - The ID of the Capacity Reservation.

$sel:sriovNetSupport:Instance', instance_sriovNetSupport - Specifies whether enhanced networking with the Intel 82599 Virtual Function interface is enabled.

$sel:publicIpAddress:Instance', instance_publicIpAddress - The public IPv4 address, or the Carrier IP address assigned to the instance, if applicable.

A Carrier IP address only applies to an instance launched in a subnet associated with a Wavelength Zone.

$sel:blockDeviceMappings:Instance', instance_blockDeviceMappings - Any block device mapping entries for the instance.

$sel:subnetId:Instance', instance_subnetId - [EC2-VPC] The ID of the subnet in which the instance is running.

$sel:enclaveOptions:Instance', instance_enclaveOptions - Indicates whether the instance is enabled for Amazon Web Services Nitro Enclaves.

$sel:kernelId:Instance', instance_kernelId - The kernel associated with this instance, if applicable.

$sel:cpuOptions:Instance', instance_cpuOptions - The CPU options for the instance.

$sel:privateDnsName:Instance', instance_privateDnsName - (IPv4 only) The private DNS hostname name assigned to the instance. This DNS hostname can only be used inside the Amazon EC2 network. This name is not available until the instance enters the running state.

EC2-VPC

The Amazon-provided DNS server resolves Amazon-provided private DNS hostnames if you've enabled DNS resolution and DNS hostnames in your VPC. If you are not using the Amazon-provided DNS server in your VPC, your custom domain name servers must resolve the hostname as appropriate.

$sel:licenses:Instance', instance_licenses - The license configurations for the instance.

$sel:vpcId:Instance', instance_vpcId - [EC2-VPC] The ID of the VPC in which the instance is running.

$sel:networkInterfaces:Instance', instance_networkInterfaces - [EC2-VPC] The network interfaces for the instance.

$sel:keyName:Instance', instance_keyName - The name of the key pair, if this instance was launched with an associated key pair.

$sel:spotInstanceRequestId:Instance', instance_spotInstanceRequestId - If the request is a Spot Instance request, the ID of the request.

$sel:elasticGpuAssociations:Instance', instance_elasticGpuAssociations - The Elastic GPU associated with the instance.

$sel:enaSupport:Instance', instance_enaSupport - Specifies whether enhanced networking with ENA is enabled.

$sel:metadataOptions:Instance', instance_metadataOptions - The metadata options for the instance.

$sel:privateIpAddress:Instance', instance_privateIpAddress - The private IPv4 address assigned to the instance.

$sel:clientToken:Instance', instance_clientToken - The idempotency token you provided when you launched the instance, if applicable.

$sel:instanceId:Instance', instance_instanceId - The ID of the instance.

$sel:imageId:Instance', instance_imageId - The ID of the AMI used to launch the instance.

$sel:amiLaunchIndex:Instance', instance_amiLaunchIndex - The AMI launch index, which can be used to find this instance in the launch group.

$sel:instanceType:Instance', instance_instanceType - The instance type.

$sel:launchTime:Instance', instance_launchTime - The time the instance was launched.

$sel:placement:Instance', instance_placement - The location where the instance launched, if applicable.

$sel:monitoring:Instance', instance_monitoring - The monitoring for the instance.

$sel:architecture:Instance', instance_architecture - The architecture of the image.

$sel:rootDeviceType:Instance', instance_rootDeviceType - The root device type used by the AMI. The AMI can use an EBS volume or an instance store volume.

$sel:virtualizationType:Instance', instance_virtualizationType - The virtualization type of the instance.

$sel:hypervisor:Instance', instance_hypervisor - The hypervisor type of the instance. The value xen is used for both Xen and Nitro hypervisors.

$sel:state:Instance', instance_state - The current state of the instance.

instance_platform :: Lens' Instance (Maybe PlatformValues)

The value is Windows for Windows instances; otherwise blank.

instance_instanceLifecycle :: Lens' Instance (Maybe InstanceLifecycleType)

Indicates whether this is a Spot Instance or a Scheduled Instance.

instance_stateTransitionReason :: Lens' Instance (Maybe Text)

The reason for the most recent state transition. This might be an empty string.

instance_capacityReservationSpecification :: Lens' Instance (Maybe CapacityReservationSpecificationResponse)

Information about the Capacity Reservation targeting option.

instance_ebsOptimized :: Lens' Instance (Maybe Bool)

Indicates whether the instance is optimized for Amazon EBS I/O. This optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal I/O performance. This optimization isn't available with all instance types. Additional usage charges apply when using an EBS Optimized instance.

instance_rootDeviceName :: Lens' Instance (Maybe Text)

The device name of the root device volume (for example, /dev/sda1).

instance_ramdiskId :: Lens' Instance (Maybe Text)

The RAM disk associated with this instance, if applicable.

instance_stateReason :: Lens' Instance (Maybe StateReason)

The reason for the most recent state transition.

instance_outpostArn :: Lens' Instance (Maybe Text)

The Amazon Resource Name (ARN) of the Outpost.

instance_elasticInferenceAcceleratorAssociations :: Lens' Instance (Maybe [ElasticInferenceAcceleratorAssociation])

The elastic inference accelerator associated with the instance.

instance_usageOperation :: Lens' Instance (Maybe Text)

The usage operation value for the instance. For more information, see AMI billing information fields in the Amazon EC2 User Guide.

instance_platformDetails :: Lens' Instance (Maybe Text)

The platform details value for the instance. For more information, see AMI billing information fields in the Amazon EC2 User Guide.

instance_bootMode :: Lens' Instance (Maybe BootModeValues)

The boot mode of the instance. For more information, see Boot modes in the Amazon EC2 User Guide.

instance_productCodes :: Lens' Instance (Maybe [ProductCode])

The product codes attached to this instance, if applicable.

instance_sourceDestCheck :: Lens' Instance (Maybe Bool)

Indicates whether source/destination checking is enabled.

instance_securityGroups :: Lens' Instance (Maybe [GroupIdentifier])

The security groups for the instance.

instance_publicDnsName :: Lens' Instance (Maybe Text)

(IPv4 only) The public DNS name assigned to the instance. This name is not available until the instance enters the running state. For EC2-VPC, this name is only available if you've enabled DNS hostnames for your VPC.

instance_iamInstanceProfile :: Lens' Instance (Maybe IamInstanceProfile)

The IAM instance profile associated with the instance, if applicable.

instance_hibernationOptions :: Lens' Instance (Maybe HibernationOptions)

Indicates whether the instance is enabled for hibernation.

instance_usageOperationUpdateTime :: Lens' Instance (Maybe UTCTime)

The time that the usage operation was last updated.

instance_tags :: Lens' Instance (Maybe [Tag])

Any tags assigned to the instance.

instance_capacityReservationId :: Lens' Instance (Maybe Text)

The ID of the Capacity Reservation.

instance_sriovNetSupport :: Lens' Instance (Maybe Text)

Specifies whether enhanced networking with the Intel 82599 Virtual Function interface is enabled.

instance_publicIpAddress :: Lens' Instance (Maybe Text)

The public IPv4 address, or the Carrier IP address assigned to the instance, if applicable.

A Carrier IP address only applies to an instance launched in a subnet associated with a Wavelength Zone.

instance_blockDeviceMappings :: Lens' Instance (Maybe [InstanceBlockDeviceMapping])

Any block device mapping entries for the instance.

instance_subnetId :: Lens' Instance (Maybe Text)

EC2-VPC

The ID of the subnet in which the instance is running.

instance_enclaveOptions :: Lens' Instance (Maybe EnclaveOptions)

Indicates whether the instance is enabled for Amazon Web Services Nitro Enclaves.

instance_kernelId :: Lens' Instance (Maybe Text)

The kernel associated with this instance, if applicable.

instance_cpuOptions :: Lens' Instance (Maybe CpuOptions)

The CPU options for the instance.

instance_privateDnsName :: Lens' Instance (Maybe Text)

(IPv4 only) The private DNS hostname name assigned to the instance. This DNS hostname can only be used inside the Amazon EC2 network. This name is not available until the instance enters the running state.

EC2-VPC

The Amazon-provided DNS server resolves Amazon-provided private DNS hostnames if you've enabled DNS resolution and DNS hostnames in your VPC. If you are not using the Amazon-provided DNS server in your VPC, your custom domain name servers must resolve the hostname as appropriate.

instance_licenses :: Lens' Instance (Maybe [LicenseConfiguration])

The license configurations for the instance.

instance_vpcId :: Lens' Instance (Maybe Text)

EC2-VPC

The ID of the VPC in which the instance is running.

instance_networkInterfaces :: Lens' Instance (Maybe [InstanceNetworkInterface])

EC2-VPC

The network interfaces for the instance.

instance_keyName :: Lens' Instance (Maybe Text)

The name of the key pair, if this instance was launched with an associated key pair.

instance_spotInstanceRequestId :: Lens' Instance (Maybe Text)

If the request is a Spot Instance request, the ID of the request.

instance_elasticGpuAssociations :: Lens' Instance (Maybe [ElasticGpuAssociation])

The Elastic GPU associated with the instance.

instance_enaSupport :: Lens' Instance (Maybe Bool)

Specifies whether enhanced networking with ENA is enabled.

instance_metadataOptions :: Lens' Instance (Maybe InstanceMetadataOptionsResponse)

The metadata options for the instance.

instance_privateIpAddress :: Lens' Instance (Maybe Text)

The private IPv4 address assigned to the instance.

instance_clientToken :: Lens' Instance (Maybe Text)

The idempotency token you provided when you launched the instance, if applicable.

instance_instanceId :: Lens' Instance Text

The ID of the instance.

instance_imageId :: Lens' Instance Text

The ID of the AMI used to launch the instance.

instance_amiLaunchIndex :: Lens' Instance Int

The AMI launch index, which can be used to find this instance in the launch group.

instance_instanceType :: Lens' Instance InstanceType

The instance type.

instance_launchTime :: Lens' Instance UTCTime

The time the instance was launched.

instance_placement :: Lens' Instance Placement

The location where the instance launched, if applicable.

instance_monitoring :: Lens' Instance Monitoring

The monitoring for the instance.

instance_architecture :: Lens' Instance ArchitectureValues

The architecture of the image.

instance_rootDeviceType :: Lens' Instance DeviceType

The root device type used by the AMI. The AMI can use an EBS volume or an instance store volume.

instance_virtualizationType :: Lens' Instance VirtualizationType

The virtualization type of the instance.

instance_hypervisor :: Lens' Instance HypervisorType

The hypervisor type of the instance. The value xen is used for both Xen and Nitro hypervisors.

instance_state :: Lens' Instance InstanceState

The current state of the instance.