Network.AWS.EC2.Types.InstanceMetadataOptionsRequest

Description

Documentation

data InstanceMetadataOptionsRequest

The metadata options for the instance.

See: newInstanceMetadataOptionsRequest smart constructor.

Constructors

InstanceMetadataOptionsRequest'

Fields

httpProtocolIpv6 :: Maybe InstanceMetadataProtocolState

Enables or disables the IPv6 endpoint for the instance metadata service.

httpEndpoint :: Maybe InstanceMetadataEndpointState

This parameter enables or disables the HTTP metadata endpoint on your instances. If the parameter is not specified, the default state is enabled.

If you specify a value of disabled, you will not be able to access your instance metadata.

httpPutResponseHopLimit :: Maybe Int

The desired HTTP PUT response hop limit for instance metadata requests. The larger the number, the further instance metadata requests can travel.

Default: 1

Possible values: Integers from 1 to 64

httpTokens :: Maybe HttpTokensState

The state of token usage for your instance metadata requests. If the parameter is not specified in the request, the default state is optional.

If the state is optional, you can choose to retrieve instance metadata with or without a signed token header on your request. If you retrieve the IAM role credentials without a token, the version 1.0 role credentials are returned. If you retrieve the IAM role credentials using a valid signed token, the version 2.0 role credentials are returned.

If the state is required, you must send a signed token header with any instance metadata retrieval requests. In this state, retrieving the IAM role credentials always returns the version 2.0 credentials; the version 1.0 credentials are not available.

newInstanceMetadataOptionsRequest :: InstanceMetadataOptionsRequest

Create a value of InstanceMetadataOptionsRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpProtocolIpv6:InstanceMetadataOptionsRequest', instanceMetadataOptionsRequest_httpProtocolIpv6 - Enables or disables the IPv6 endpoint for the instance metadata service.

$sel:httpEndpoint:InstanceMetadataOptionsRequest', instanceMetadataOptionsRequest_httpEndpoint - This parameter enables or disables the HTTP metadata endpoint on your instances. If the parameter is not specified, the default state is enabled.

If you specify a value of disabled, you will not be able to access your instance metadata.

$sel:httpPutResponseHopLimit:InstanceMetadataOptionsRequest', instanceMetadataOptionsRequest_httpPutResponseHopLimit - The desired HTTP PUT response hop limit for instance metadata requests. The larger the number, the further instance metadata requests can travel.

Default: 1

Possible values: Integers from 1 to 64

$sel:httpTokens:InstanceMetadataOptionsRequest', instanceMetadataOptionsRequest_httpTokens - The state of token usage for your instance metadata requests. If the parameter is not specified in the request, the default state is optional.

If the state is optional, you can choose to retrieve instance metadata with or without a signed token header on your request. If you retrieve the IAM role credentials without a token, the version 1.0 role credentials are returned. If you retrieve the IAM role credentials using a valid signed token, the version 2.0 role credentials are returned.

If the state is required, you must send a signed token header with any instance metadata retrieval requests. In this state, retrieving the IAM role credentials always returns the version 2.0 credentials; the version 1.0 credentials are not available.

instanceMetadataOptionsRequest_httpProtocolIpv6 :: Lens' InstanceMetadataOptionsRequest (Maybe InstanceMetadataProtocolState)

Enables or disables the IPv6 endpoint for the instance metadata service.

instanceMetadataOptionsRequest_httpEndpoint :: Lens' InstanceMetadataOptionsRequest (Maybe InstanceMetadataEndpointState)

This parameter enables or disables the HTTP metadata endpoint on your instances. If the parameter is not specified, the default state is enabled.

If you specify a value of disabled, you will not be able to access your instance metadata.

instanceMetadataOptionsRequest_httpPutResponseHopLimit :: Lens' InstanceMetadataOptionsRequest (Maybe Int)

The desired HTTP PUT response hop limit for instance metadata requests. The larger the number, the further instance metadata requests can travel.

Default: 1

Possible values: Integers from 1 to 64

instanceMetadataOptionsRequest_httpTokens :: Lens' InstanceMetadataOptionsRequest (Maybe HttpTokensState)

The state of token usage for your instance metadata requests. If the parameter is not specified in the request, the default state is optional.

If the state is optional, you can choose to retrieve instance metadata with or without a signed token header on your request. If you retrieve the IAM role credentials without a token, the version 1.0 role credentials are returned. If you retrieve the IAM role credentials using a valid signed token, the version 2.0 role credentials are returned.

If the state is required, you must send a signed token header with any instance metadata retrieval requests. In this state, retrieving the IAM role credentials always returns the version 2.0 credentials; the version 1.0 credentials are not available.