Network.AWS.EC2.Types.InstanceMetadataOptionsResponse

Description

Documentation

data InstanceMetadataOptionsResponse

The metadata options for the instance.

See: newInstanceMetadataOptionsResponse smart constructor.

Constructors

InstanceMetadataOptionsResponse'

Fields

httpProtocolIpv6 :: Maybe InstanceMetadataProtocolState

Whether or not the IPv6 endpoint for the instance metadata service is enabled or disabled.

httpEndpoint :: Maybe InstanceMetadataEndpointState

This parameter enables or disables the HTTP metadata endpoint on your instances. If the parameter is not specified, the default state is enabled.

If you specify a value of disabled, you will not be able to access your instance metadata.

httpPutResponseHopLimit :: Maybe Int

The desired HTTP PUT response hop limit for instance metadata requests. The larger the number, the further instance metadata requests can travel.

Default: 1

Possible values: Integers from 1 to 64

state :: Maybe InstanceMetadataOptionsState

The state of the metadata option changes.

pending - The metadata options are being updated and the instance is not ready to process metadata traffic with the new selection.

applied - The metadata options have been successfully applied on the instance.

httpTokens :: Maybe HttpTokensState

The state of token usage for your instance metadata requests. If the parameter is not specified in the request, the default state is optional.

If the state is optional, you can choose to retrieve instance metadata with or without a signed token header on your request. If you retrieve the IAM role credentials without a token, the version 1.0 role credentials are returned. If you retrieve the IAM role credentials using a valid signed token, the version 2.0 role credentials are returned.

If the state is required, you must send a signed token header with any instance metadata retrieval requests. In this state, retrieving the IAM role credential always returns the version 2.0 credentials; the version 1.0 credentials are not available.

newInstanceMetadataOptionsResponse :: InstanceMetadataOptionsResponse

Create a value of InstanceMetadataOptionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpProtocolIpv6:InstanceMetadataOptionsResponse', instanceMetadataOptionsResponse_httpProtocolIpv6 - Whether or not the IPv6 endpoint for the instance metadata service is enabled or disabled.

$sel:httpEndpoint:InstanceMetadataOptionsResponse', instanceMetadataOptionsResponse_httpEndpoint - This parameter enables or disables the HTTP metadata endpoint on your instances. If the parameter is not specified, the default state is enabled.

If you specify a value of disabled, you will not be able to access your instance metadata.

$sel:httpPutResponseHopLimit:InstanceMetadataOptionsResponse', instanceMetadataOptionsResponse_httpPutResponseHopLimit - The desired HTTP PUT response hop limit for instance metadata requests. The larger the number, the further instance metadata requests can travel.

Default: 1

Possible values: Integers from 1 to 64

$sel:state:InstanceMetadataOptionsResponse', instanceMetadataOptionsResponse_state - The state of the metadata option changes.

pending - The metadata options are being updated and the instance is not ready to process metadata traffic with the new selection.

applied - The metadata options have been successfully applied on the instance.

$sel:httpTokens:InstanceMetadataOptionsResponse', instanceMetadataOptionsResponse_httpTokens - The state of token usage for your instance metadata requests. If the parameter is not specified in the request, the default state is optional.

If the state is optional, you can choose to retrieve instance metadata with or without a signed token header on your request. If you retrieve the IAM role credentials without a token, the version 1.0 role credentials are returned. If you retrieve the IAM role credentials using a valid signed token, the version 2.0 role credentials are returned.

If the state is required, you must send a signed token header with any instance metadata retrieval requests. In this state, retrieving the IAM role credential always returns the version 2.0 credentials; the version 1.0 credentials are not available.

instanceMetadataOptionsResponse_httpProtocolIpv6 :: Lens' InstanceMetadataOptionsResponse (Maybe InstanceMetadataProtocolState)

Whether or not the IPv6 endpoint for the instance metadata service is enabled or disabled.

instanceMetadataOptionsResponse_httpEndpoint :: Lens' InstanceMetadataOptionsResponse (Maybe InstanceMetadataEndpointState)

This parameter enables or disables the HTTP metadata endpoint on your instances. If the parameter is not specified, the default state is enabled.

If you specify a value of disabled, you will not be able to access your instance metadata.

instanceMetadataOptionsResponse_httpPutResponseHopLimit :: Lens' InstanceMetadataOptionsResponse (Maybe Int)

The desired HTTP PUT response hop limit for instance metadata requests. The larger the number, the further instance metadata requests can travel.

Default: 1

Possible values: Integers from 1 to 64

instanceMetadataOptionsResponse_state :: Lens' InstanceMetadataOptionsResponse (Maybe InstanceMetadataOptionsState)

The state of the metadata option changes.

pending - The metadata options are being updated and the instance is not ready to process metadata traffic with the new selection.

applied - The metadata options have been successfully applied on the instance.

instanceMetadataOptionsResponse_httpTokens :: Lens' InstanceMetadataOptionsResponse (Maybe HttpTokensState)

The state of token usage for your instance metadata requests. If the parameter is not specified in the request, the default state is optional.

If the state is optional, you can choose to retrieve instance metadata with or without a signed token header on your request. If you retrieve the IAM role credentials without a token, the version 1.0 role credentials are returned. If you retrieve the IAM role credentials using a valid signed token, the version 2.0 role credentials are returned.

If the state is required, you must send a signed token header with any instance metadata retrieval requests. In this state, retrieving the IAM role credential always returns the version 2.0 credentials; the version 1.0 credentials are not available.