Network.AWS.EC2.Types.InstanceNetworkInterface

Description

Documentation

data InstanceNetworkInterface

Describes a network interface.

See: newInstanceNetworkInterface smart constructor.

Constructors

InstanceNetworkInterface'

Fields

status :: Maybe NetworkInterfaceStatus

The status of the network interface.

groups :: Maybe [GroupIdentifier]

One or more security groups.

ownerId :: Maybe Text

The ID of the Amazon Web Services account that created the network interface.

privateIpAddresses :: Maybe [InstancePrivateIpAddress]

One or more private IPv4 addresses associated with the network interface.

macAddress :: Maybe Text

The MAC address.

attachment :: Maybe InstanceNetworkInterfaceAttachment

The network interface attachment.

association :: Maybe InstanceNetworkInterfaceAssociation

The association information for an Elastic IPv4 associated with the network interface.

ipv6Addresses :: Maybe [InstanceIpv6Address]

One or more IPv6 addresses associated with the network interface.

sourceDestCheck :: Maybe Bool

Indicates whether source/destination checking is enabled.

interfaceType :: Maybe Text

Describes the type of network interface.

Valid values: interface | efa | trunk

ipv6Prefixes :: Maybe [InstanceIpv6Prefix]

The IPv6 delegated prefixes that are assigned to the network interface.

description :: Maybe Text

The description.

subnetId :: Maybe Text

The ID of the subnet.

networkInterfaceId :: Maybe Text

The ID of the network interface.

privateDnsName :: Maybe Text

The private DNS name.

vpcId :: Maybe Text

The ID of the VPC.

ipv4Prefixes :: Maybe [InstanceIpv4Prefix]

The IPv4 delegated prefixes that are assigned to the network interface.

privateIpAddress :: Maybe Text

The IPv4 address of the network interface within the subnet.

newInstanceNetworkInterface :: InstanceNetworkInterface

Create a value of InstanceNetworkInterface with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:InstanceNetworkInterface', instanceNetworkInterface_status - The status of the network interface.

$sel:groups:InstanceNetworkInterface', instanceNetworkInterface_groups - One or more security groups.

$sel:ownerId:InstanceNetworkInterface', instanceNetworkInterface_ownerId - The ID of the Amazon Web Services account that created the network interface.

$sel:privateIpAddresses:InstanceNetworkInterface', instanceNetworkInterface_privateIpAddresses - One or more private IPv4 addresses associated with the network interface.

$sel:macAddress:InstanceNetworkInterface', instanceNetworkInterface_macAddress - The MAC address.

$sel:attachment:InstanceNetworkInterface', instanceNetworkInterface_attachment - The network interface attachment.

$sel:association:InstanceNetworkInterface', instanceNetworkInterface_association - The association information for an Elastic IPv4 associated with the network interface.

$sel:ipv6Addresses:InstanceNetworkInterface', instanceNetworkInterface_ipv6Addresses - One or more IPv6 addresses associated with the network interface.

$sel:sourceDestCheck:InstanceNetworkInterface', instanceNetworkInterface_sourceDestCheck - Indicates whether source/destination checking is enabled.

$sel:interfaceType:InstanceNetworkInterface', instanceNetworkInterface_interfaceType - Describes the type of network interface.

Valid values: interface | efa | trunk

$sel:ipv6Prefixes:InstanceNetworkInterface', instanceNetworkInterface_ipv6Prefixes - The IPv6 delegated prefixes that are assigned to the network interface.

$sel:description:InstanceNetworkInterface', instanceNetworkInterface_description - The description.

$sel:subnetId:InstanceNetworkInterface', instanceNetworkInterface_subnetId - The ID of the subnet.

$sel:networkInterfaceId:InstanceNetworkInterface', instanceNetworkInterface_networkInterfaceId - The ID of the network interface.

$sel:privateDnsName:InstanceNetworkInterface', instanceNetworkInterface_privateDnsName - The private DNS name.

$sel:vpcId:InstanceNetworkInterface', instanceNetworkInterface_vpcId - The ID of the VPC.

$sel:ipv4Prefixes:InstanceNetworkInterface', instanceNetworkInterface_ipv4Prefixes - The IPv4 delegated prefixes that are assigned to the network interface.

$sel:privateIpAddress:InstanceNetworkInterface', instanceNetworkInterface_privateIpAddress - The IPv4 address of the network interface within the subnet.

instanceNetworkInterface_status :: Lens' InstanceNetworkInterface (Maybe NetworkInterfaceStatus)

The status of the network interface.

instanceNetworkInterface_groups :: Lens' InstanceNetworkInterface (Maybe [GroupIdentifier])

One or more security groups.

instanceNetworkInterface_ownerId :: Lens' InstanceNetworkInterface (Maybe Text)

The ID of the Amazon Web Services account that created the network interface.

instanceNetworkInterface_privateIpAddresses :: Lens' InstanceNetworkInterface (Maybe [InstancePrivateIpAddress])

One or more private IPv4 addresses associated with the network interface.

instanceNetworkInterface_macAddress :: Lens' InstanceNetworkInterface (Maybe Text)

The MAC address.

instanceNetworkInterface_attachment :: Lens' InstanceNetworkInterface (Maybe InstanceNetworkInterfaceAttachment)

The network interface attachment.

instanceNetworkInterface_association :: Lens' InstanceNetworkInterface (Maybe InstanceNetworkInterfaceAssociation)

The association information for an Elastic IPv4 associated with the network interface.

instanceNetworkInterface_ipv6Addresses :: Lens' InstanceNetworkInterface (Maybe [InstanceIpv6Address])

One or more IPv6 addresses associated with the network interface.

instanceNetworkInterface_sourceDestCheck :: Lens' InstanceNetworkInterface (Maybe Bool)

Indicates whether source/destination checking is enabled.

instanceNetworkInterface_interfaceType :: Lens' InstanceNetworkInterface (Maybe Text)

Describes the type of network interface.

Valid values: interface | efa | trunk

instanceNetworkInterface_ipv6Prefixes :: Lens' InstanceNetworkInterface (Maybe [InstanceIpv6Prefix])

The IPv6 delegated prefixes that are assigned to the network interface.

instanceNetworkInterface_description :: Lens' InstanceNetworkInterface (Maybe Text)

The description.

instanceNetworkInterface_subnetId :: Lens' InstanceNetworkInterface (Maybe Text)

The ID of the subnet.

instanceNetworkInterface_networkInterfaceId :: Lens' InstanceNetworkInterface (Maybe Text)

The ID of the network interface.

instanceNetworkInterface_privateDnsName :: Lens' InstanceNetworkInterface (Maybe Text)

The private DNS name.

instanceNetworkInterface_vpcId :: Lens' InstanceNetworkInterface (Maybe Text)

The ID of the VPC.

instanceNetworkInterface_ipv4Prefixes :: Lens' InstanceNetworkInterface (Maybe [InstanceIpv4Prefix])

The IPv4 delegated prefixes that are assigned to the network interface.

instanceNetworkInterface_privateIpAddress :: Lens' InstanceNetworkInterface (Maybe Text)

The IPv4 address of the network interface within the subnet.