Network.AWS.EC2.Types.InstanceNetworkInterfaceSpecification

Description

Documentation

data InstanceNetworkInterfaceSpecification

Describes a network interface.

See: newInstanceNetworkInterfaceSpecification smart constructor.

Constructors

InstanceNetworkInterfaceSpecification'

Fields

groups :: Maybe [Text]

The IDs of the security groups for the network interface. Applies only if creating a network interface when launching an instance.

privateIpAddresses :: Maybe [PrivateIpAddressSpecification]

One or more private IPv4 addresses to assign to the network interface. Only one private IPv4 address can be designated as primary. You cannot specify this option if you're launching more than one instance in a RunInstances request.

deleteOnTermination :: Maybe Bool

If set to true, the interface is deleted when the instance is terminated. You can specify true only if creating a new network interface when launching an instance.

associateCarrierIpAddress :: Maybe Bool

Indicates whether to assign a carrier IP address to the network interface.

You can only assign a carrier IP address to a network interface that is in a subnet in a Wavelength Zone. For more information about carrier IP addresses, see Carrier IP addresses in the Amazon Web Services Wavelength Developer Guide.

associatePublicIpAddress :: Maybe Bool

Indicates whether to assign a public IPv4 address to an instance you launch in a VPC. The public IP address can only be assigned to a network interface for eth0, and can only be assigned to a new network interface, not an existing one. You cannot specify more than one network interface in the request. If launching into a default subnet, the default value is true.

ipv4Prefixes :: Maybe [Ipv4PrefixSpecificationRequest]

One or more IPv4 delegated prefixes to be assigned to the network interface. You cannot use this option if you use the Ipv4PrefixCount option.

interfaceType :: Maybe Text

The type of network interface.

To create an Elastic Fabric Adapter (EFA), specify efa. For more information, see Elastic Fabric Adapter in the Amazon Elastic Compute Cloud User Guide.

Valid values: interface | efa

ipv4PrefixCount :: Maybe Int

The number of IPv4 delegated prefixes to be automatically assigned to the network interface. You cannot use this option if you use the Ipv4Prefix option.

networkInterfaceId :: Maybe Text

The ID of the network interface.

If you are creating a Spot Fleet, omit this parameter because you can’t specify a network interface ID in a launch specification.

subnetId :: Maybe Text

The ID of the subnet associated with the network interface. Applies only if creating a network interface when launching an instance.

ipv6AddressCount :: Maybe Int

A number of IPv6 addresses to assign to the network interface. Amazon EC2 chooses the IPv6 addresses from the range of the subnet. You cannot specify this option and the option to assign specific IPv6 addresses in the same request. You can specify this option if you've specified a minimum number of instances to launch.

networkCardIndex :: Maybe Int

The index of the network card. Some instance types support multiple network cards. The primary network interface must be assigned to network card index 0. The default is network card index 0.

If you are using RequestSpotInstances to create Spot Instances, omit this parameter because you can’t specify the network card index when using this API. To specify the network card index, use RunInstances.

ipv6Prefixes :: Maybe [Ipv6PrefixSpecificationRequest]

One or more IPv6 delegated prefixes to be assigned to the network interface. You cannot use this option if you use the Ipv6PrefixCount option.

privateIpAddress :: Maybe Text

The private IPv4 address of the network interface. Applies only if creating a network interface when launching an instance. You cannot specify this option if you're launching more than one instance in a RunInstances request.

ipv6PrefixCount :: Maybe Int

The number of IPv6 delegated prefixes to be automatically assigned to the network interface. You cannot use this option if you use the Ipv6Prefix option.

secondaryPrivateIpAddressCount :: Maybe Int

The number of secondary private IPv4 addresses. You can't specify this option and specify more than one private IP address using the private IP addresses option. You cannot specify this option if you're launching more than one instance in a RunInstances request.

description :: Maybe Text

The description of the network interface. Applies only if creating a network interface when launching an instance.

deviceIndex :: Maybe Int

The position of the network interface in the attachment order. A primary network interface has a device index of 0.

If you specify a network interface when launching an instance, you must specify the device index.

ipv6Addresses :: Maybe [InstanceIpv6Address]

One or more IPv6 addresses to assign to the network interface. You cannot specify this option and the option to assign a number of IPv6 addresses in the same request. You cannot specify this option if you've specified a minimum number of instances to launch.

newInstanceNetworkInterfaceSpecification :: InstanceNetworkInterfaceSpecification

Create a value of InstanceNetworkInterfaceSpecification with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groups:InstanceNetworkInterfaceSpecification', instanceNetworkInterfaceSpecification_groups - The IDs of the security groups for the network interface. Applies only if creating a network interface when launching an instance.

$sel:privateIpAddresses:InstanceNetworkInterfaceSpecification', instanceNetworkInterfaceSpecification_privateIpAddresses - One or more private IPv4 addresses to assign to the network interface. Only one private IPv4 address can be designated as primary. You cannot specify this option if you're launching more than one instance in a RunInstances request.

$sel:deleteOnTermination:InstanceNetworkInterfaceSpecification', instanceNetworkInterfaceSpecification_deleteOnTermination - If set to true, the interface is deleted when the instance is terminated. You can specify true only if creating a new network interface when launching an instance.

$sel:associateCarrierIpAddress:InstanceNetworkInterfaceSpecification', instanceNetworkInterfaceSpecification_associateCarrierIpAddress - Indicates whether to assign a carrier IP address to the network interface.

You can only assign a carrier IP address to a network interface that is in a subnet in a Wavelength Zone. For more information about carrier IP addresses, see Carrier IP addresses in the Amazon Web Services Wavelength Developer Guide.

$sel:associatePublicIpAddress:InstanceNetworkInterfaceSpecification', instanceNetworkInterfaceSpecification_associatePublicIpAddress - Indicates whether to assign a public IPv4 address to an instance you launch in a VPC. The public IP address can only be assigned to a network interface for eth0, and can only be assigned to a new network interface, not an existing one. You cannot specify more than one network interface in the request. If launching into a default subnet, the default value is true.

$sel:ipv4Prefixes:InstanceNetworkInterfaceSpecification', instanceNetworkInterfaceSpecification_ipv4Prefixes - One or more IPv4 delegated prefixes to be assigned to the network interface. You cannot use this option if you use the Ipv4PrefixCount option.

$sel:interfaceType:InstanceNetworkInterfaceSpecification', instanceNetworkInterfaceSpecification_interfaceType - The type of network interface.

To create an Elastic Fabric Adapter (EFA), specify efa. For more information, see Elastic Fabric Adapter in the Amazon Elastic Compute Cloud User Guide.

Valid values: interface | efa

$sel:ipv4PrefixCount:InstanceNetworkInterfaceSpecification', instanceNetworkInterfaceSpecification_ipv4PrefixCount - The number of IPv4 delegated prefixes to be automatically assigned to the network interface. You cannot use this option if you use the Ipv4Prefix option.

$sel:networkInterfaceId:InstanceNetworkInterfaceSpecification', instanceNetworkInterfaceSpecification_networkInterfaceId - The ID of the network interface.

If you are creating a Spot Fleet, omit this parameter because you can’t specify a network interface ID in a launch specification.

$sel:subnetId:InstanceNetworkInterfaceSpecification', instanceNetworkInterfaceSpecification_subnetId - The ID of the subnet associated with the network interface. Applies only if creating a network interface when launching an instance.

$sel:ipv6AddressCount:InstanceNetworkInterfaceSpecification', instanceNetworkInterfaceSpecification_ipv6AddressCount - A number of IPv6 addresses to assign to the network interface. Amazon EC2 chooses the IPv6 addresses from the range of the subnet. You cannot specify this option and the option to assign specific IPv6 addresses in the same request. You can specify this option if you've specified a minimum number of instances to launch.

$sel:networkCardIndex:InstanceNetworkInterfaceSpecification', instanceNetworkInterfaceSpecification_networkCardIndex - The index of the network card. Some instance types support multiple network cards. The primary network interface must be assigned to network card index 0. The default is network card index 0.

If you are using RequestSpotInstances to create Spot Instances, omit this parameter because you can’t specify the network card index when using this API. To specify the network card index, use RunInstances.

$sel:ipv6Prefixes:InstanceNetworkInterfaceSpecification', instanceNetworkInterfaceSpecification_ipv6Prefixes - One or more IPv6 delegated prefixes to be assigned to the network interface. You cannot use this option if you use the Ipv6PrefixCount option.

$sel:privateIpAddress:InstanceNetworkInterfaceSpecification', instanceNetworkInterfaceSpecification_privateIpAddress - The private IPv4 address of the network interface. Applies only if creating a network interface when launching an instance. You cannot specify this option if you're launching more than one instance in a RunInstances request.

$sel:ipv6PrefixCount:InstanceNetworkInterfaceSpecification', instanceNetworkInterfaceSpecification_ipv6PrefixCount - The number of IPv6 delegated prefixes to be automatically assigned to the network interface. You cannot use this option if you use the Ipv6Prefix option.

$sel:secondaryPrivateIpAddressCount:InstanceNetworkInterfaceSpecification', instanceNetworkInterfaceSpecification_secondaryPrivateIpAddressCount - The number of secondary private IPv4 addresses. You can't specify this option and specify more than one private IP address using the private IP addresses option. You cannot specify this option if you're launching more than one instance in a RunInstances request.

$sel:description:InstanceNetworkInterfaceSpecification', instanceNetworkInterfaceSpecification_description - The description of the network interface. Applies only if creating a network interface when launching an instance.

$sel:deviceIndex:InstanceNetworkInterfaceSpecification', instanceNetworkInterfaceSpecification_deviceIndex - The position of the network interface in the attachment order. A primary network interface has a device index of 0.

If you specify a network interface when launching an instance, you must specify the device index.

$sel:ipv6Addresses:InstanceNetworkInterfaceSpecification', instanceNetworkInterfaceSpecification_ipv6Addresses - One or more IPv6 addresses to assign to the network interface. You cannot specify this option and the option to assign a number of IPv6 addresses in the same request. You cannot specify this option if you've specified a minimum number of instances to launch.

instanceNetworkInterfaceSpecification_groups :: Lens' InstanceNetworkInterfaceSpecification (Maybe [Text])

The IDs of the security groups for the network interface. Applies only if creating a network interface when launching an instance.

instanceNetworkInterfaceSpecification_privateIpAddresses :: Lens' InstanceNetworkInterfaceSpecification (Maybe [PrivateIpAddressSpecification])

One or more private IPv4 addresses to assign to the network interface. Only one private IPv4 address can be designated as primary. You cannot specify this option if you're launching more than one instance in a RunInstances request.

instanceNetworkInterfaceSpecification_deleteOnTermination :: Lens' InstanceNetworkInterfaceSpecification (Maybe Bool)

If set to true, the interface is deleted when the instance is terminated. You can specify true only if creating a new network interface when launching an instance.

instanceNetworkInterfaceSpecification_associateCarrierIpAddress :: Lens' InstanceNetworkInterfaceSpecification (Maybe Bool)

Indicates whether to assign a carrier IP address to the network interface.

You can only assign a carrier IP address to a network interface that is in a subnet in a Wavelength Zone. For more information about carrier IP addresses, see Carrier IP addresses in the Amazon Web Services Wavelength Developer Guide.

instanceNetworkInterfaceSpecification_associatePublicIpAddress :: Lens' InstanceNetworkInterfaceSpecification (Maybe Bool)

Indicates whether to assign a public IPv4 address to an instance you launch in a VPC. The public IP address can only be assigned to a network interface for eth0, and can only be assigned to a new network interface, not an existing one. You cannot specify more than one network interface in the request. If launching into a default subnet, the default value is true.

instanceNetworkInterfaceSpecification_ipv4Prefixes :: Lens' InstanceNetworkInterfaceSpecification (Maybe [Ipv4PrefixSpecificationRequest])

One or more IPv4 delegated prefixes to be assigned to the network interface. You cannot use this option if you use the Ipv4PrefixCount option.

instanceNetworkInterfaceSpecification_interfaceType :: Lens' InstanceNetworkInterfaceSpecification (Maybe Text)

The type of network interface.

To create an Elastic Fabric Adapter (EFA), specify efa. For more information, see Elastic Fabric Adapter in the Amazon Elastic Compute Cloud User Guide.

Valid values: interface | efa

instanceNetworkInterfaceSpecification_ipv4PrefixCount :: Lens' InstanceNetworkInterfaceSpecification (Maybe Int)

The number of IPv4 delegated prefixes to be automatically assigned to the network interface. You cannot use this option if you use the Ipv4Prefix option.

instanceNetworkInterfaceSpecification_networkInterfaceId :: Lens' InstanceNetworkInterfaceSpecification (Maybe Text)

The ID of the network interface.

If you are creating a Spot Fleet, omit this parameter because you can’t specify a network interface ID in a launch specification.

instanceNetworkInterfaceSpecification_subnetId :: Lens' InstanceNetworkInterfaceSpecification (Maybe Text)

The ID of the subnet associated with the network interface. Applies only if creating a network interface when launching an instance.

instanceNetworkInterfaceSpecification_ipv6AddressCount :: Lens' InstanceNetworkInterfaceSpecification (Maybe Int)

A number of IPv6 addresses to assign to the network interface. Amazon EC2 chooses the IPv6 addresses from the range of the subnet. You cannot specify this option and the option to assign specific IPv6 addresses in the same request. You can specify this option if you've specified a minimum number of instances to launch.

instanceNetworkInterfaceSpecification_networkCardIndex :: Lens' InstanceNetworkInterfaceSpecification (Maybe Int)

The index of the network card. Some instance types support multiple network cards. The primary network interface must be assigned to network card index 0. The default is network card index 0.

If you are using RequestSpotInstances to create Spot Instances, omit this parameter because you can’t specify the network card index when using this API. To specify the network card index, use RunInstances.

instanceNetworkInterfaceSpecification_ipv6Prefixes :: Lens' InstanceNetworkInterfaceSpecification (Maybe [Ipv6PrefixSpecificationRequest])

One or more IPv6 delegated prefixes to be assigned to the network interface. You cannot use this option if you use the Ipv6PrefixCount option.

instanceNetworkInterfaceSpecification_privateIpAddress :: Lens' InstanceNetworkInterfaceSpecification (Maybe Text)

The private IPv4 address of the network interface. Applies only if creating a network interface when launching an instance. You cannot specify this option if you're launching more than one instance in a RunInstances request.

instanceNetworkInterfaceSpecification_ipv6PrefixCount :: Lens' InstanceNetworkInterfaceSpecification (Maybe Int)

The number of IPv6 delegated prefixes to be automatically assigned to the network interface. You cannot use this option if you use the Ipv6Prefix option.

instanceNetworkInterfaceSpecification_secondaryPrivateIpAddressCount :: Lens' InstanceNetworkInterfaceSpecification (Maybe Int)

The number of secondary private IPv4 addresses. You can't specify this option and specify more than one private IP address using the private IP addresses option. You cannot specify this option if you're launching more than one instance in a RunInstances request.

instanceNetworkInterfaceSpecification_description :: Lens' InstanceNetworkInterfaceSpecification (Maybe Text)

The description of the network interface. Applies only if creating a network interface when launching an instance.

instanceNetworkInterfaceSpecification_deviceIndex :: Lens' InstanceNetworkInterfaceSpecification (Maybe Int)

The position of the network interface in the attachment order. A primary network interface has a device index of 0.

If you specify a network interface when launching an instance, you must specify the device index.

instanceNetworkInterfaceSpecification_ipv6Addresses :: Lens' InstanceNetworkInterfaceSpecification (Maybe [InstanceIpv6Address])

One or more IPv6 addresses to assign to the network interface. You cannot specify this option and the option to assign a number of IPv6 addresses in the same request. You cannot specify this option if you've specified a minimum number of instances to launch.