Network.AWS.EC2.Types.InstanceStatus

Description

Documentation

data InstanceStatus

Describes the status of an instance.

See: newInstanceStatus smart constructor.

Constructors

InstanceStatus'

Fields

instanceId :: Maybe Text

The ID of the instance.

outpostArn :: Maybe Text

The Amazon Resource Name (ARN) of the Outpost.

systemStatus :: Maybe InstanceStatusSummary

Reports impaired functionality that stems from issues related to the systems that support an instance, such as hardware failures and network connectivity problems.

events :: Maybe [InstanceStatusEvent]

Any scheduled events associated with the instance.

availabilityZone :: Maybe Text

The Availability Zone of the instance.

instanceStatus :: Maybe InstanceStatusSummary

Reports impaired functionality that stems from issues internal to the instance, such as impaired reachability.

instanceState :: Maybe InstanceState

The intended state of the instance. DescribeInstanceStatus requires that an instance be in the running state.

newInstanceStatus :: InstanceStatus

Create a value of InstanceStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceId:InstanceStatus', instanceStatus_instanceId - The ID of the instance.

$sel:outpostArn:InstanceStatus', instanceStatus_outpostArn - The Amazon Resource Name (ARN) of the Outpost.

$sel:systemStatus:InstanceStatus', instanceStatus_systemStatus - Reports impaired functionality that stems from issues related to the systems that support an instance, such as hardware failures and network connectivity problems.

$sel:events:InstanceStatus', instanceStatus_events - Any scheduled events associated with the instance.

$sel:availabilityZone:InstanceStatus', instanceStatus_availabilityZone - The Availability Zone of the instance.

$sel:instanceStatus:InstanceStatus', instanceStatus_instanceStatus - Reports impaired functionality that stems from issues internal to the instance, such as impaired reachability.

$sel:instanceState:InstanceStatus', instanceStatus_instanceState - The intended state of the instance. DescribeInstanceStatus requires that an instance be in the running state.

instanceStatus_instanceId :: Lens' InstanceStatus (Maybe Text)

The ID of the instance.

instanceStatus_outpostArn :: Lens' InstanceStatus (Maybe Text)

The Amazon Resource Name (ARN) of the Outpost.

instanceStatus_systemStatus :: Lens' InstanceStatus (Maybe InstanceStatusSummary)

Reports impaired functionality that stems from issues related to the systems that support an instance, such as hardware failures and network connectivity problems.

instanceStatus_events :: Lens' InstanceStatus (Maybe [InstanceStatusEvent])

Any scheduled events associated with the instance.

instanceStatus_availabilityZone :: Lens' InstanceStatus (Maybe Text)

The Availability Zone of the instance.

instanceStatus_instanceStatus :: Lens' InstanceStatus (Maybe InstanceStatusSummary)

Reports impaired functionality that stems from issues internal to the instance, such as impaired reachability.

instanceStatus_instanceState :: Lens' InstanceStatus (Maybe InstanceState)

The intended state of the instance. DescribeInstanceStatus requires that an instance be in the running state.