Network.AWS.EC2.Types.InstanceStorageInfo

Description

Documentation

data InstanceStorageInfo

Describes the instance store features that are supported by the instance type.

See: newInstanceStorageInfo smart constructor.

Constructors

InstanceStorageInfo'

Fields

totalSizeInGB :: Maybe Integer

The total size of the disks, in GB.

encryptionSupport :: Maybe InstanceStorageEncryptionSupport

Indicates whether data is encrypted at rest.

nvmeSupport :: Maybe EphemeralNvmeSupport

Indicates whether non-volatile memory express (NVMe) is supported.

disks :: Maybe [DiskInfo]

Describes the disks that are available for the instance type.

newInstanceStorageInfo :: InstanceStorageInfo

Create a value of InstanceStorageInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:totalSizeInGB:InstanceStorageInfo', instanceStorageInfo_totalSizeInGB - The total size of the disks, in GB.

$sel:encryptionSupport:InstanceStorageInfo', instanceStorageInfo_encryptionSupport - Indicates whether data is encrypted at rest.

$sel:nvmeSupport:InstanceStorageInfo', instanceStorageInfo_nvmeSupport - Indicates whether non-volatile memory express (NVMe) is supported.

$sel:disks:InstanceStorageInfo', instanceStorageInfo_disks - Describes the disks that are available for the instance type.

instanceStorageInfo_totalSizeInGB :: Lens' InstanceStorageInfo (Maybe Integer)

The total size of the disks, in GB.

instanceStorageInfo_encryptionSupport :: Lens' InstanceStorageInfo (Maybe InstanceStorageEncryptionSupport)

Indicates whether data is encrypted at rest.

instanceStorageInfo_nvmeSupport :: Lens' InstanceStorageInfo (Maybe EphemeralNvmeSupport)

Indicates whether non-volatile memory express (NVMe) is supported.

instanceStorageInfo_disks :: Lens' InstanceStorageInfo (Maybe [DiskInfo])

Describes the disks that are available for the instance type.