Network.AWS.EC2.Types.IpPermission

Description

Documentation

data IpPermission

Describes a set of permissions for a security group rule.

See: newIpPermission smart constructor.

Constructors

IpPermission'

Fields

fromPort :: Maybe Int

The start of port range for the TCP and UDP protocols, or an ICMP/ICMPv6 type number. A value of -1 indicates all ICMP/ICMPv6 types. If you specify all ICMP/ICMPv6 types, you must specify all codes.

userIdGroupPairs :: Maybe [UserIdGroupPair]

The security group and Amazon Web Services account ID pairs.

prefixListIds :: Maybe [PrefixListId]

VPC only

The prefix list IDs.

toPort :: Maybe Int

The end of port range for the TCP and UDP protocols, or an ICMP/ICMPv6 code. A value of -1 indicates all ICMP/ICMPv6 codes. If you specify all ICMP/ICMPv6 types, you must specify all codes.

ipv6Ranges :: Maybe [Ipv6Range]

VPC only

The IPv6 ranges.

ipRanges :: Maybe [IpRange]

The IPv4 ranges.

ipProtocol :: Text

The IP protocol name (tcp, udp, icmp, icmpv6) or number (see Protocol Numbers).

VPC only

Use -1 to specify all protocols. When authorizing security group rules, specifying -1 or a protocol number other than tcp, udp, icmp, or icmpv6 allows traffic on all ports, regardless of any port range you specify. For tcp, udp, and icmp, you must specify a port range. For icmpv6, the port range is optional; if you omit the port range, traffic for all types and codes is allowed.

newIpPermission

Arguments

:: Text
$sel:ipProtocol:IpPermission'

-> IpPermission

Create a value of IpPermission with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fromPort:IpPermission', ipPermission_fromPort - The start of port range for the TCP and UDP protocols, or an ICMP/ICMPv6 type number. A value of -1 indicates all ICMP/ICMPv6 types. If you specify all ICMP/ICMPv6 types, you must specify all codes.

$sel:userIdGroupPairs:IpPermission', ipPermission_userIdGroupPairs - The security group and Amazon Web Services account ID pairs.

$sel:prefixListIds:IpPermission', ipPermission_prefixListIds - [VPC only] The prefix list IDs.

$sel:toPort:IpPermission', ipPermission_toPort - The end of port range for the TCP and UDP protocols, or an ICMP/ICMPv6 code. A value of -1 indicates all ICMP/ICMPv6 codes. If you specify all ICMP/ICMPv6 types, you must specify all codes.

$sel:ipv6Ranges:IpPermission', ipPermission_ipv6Ranges - [VPC only] The IPv6 ranges.

$sel:ipRanges:IpPermission', ipPermission_ipRanges - The IPv4 ranges.

$sel:ipProtocol:IpPermission', ipPermission_ipProtocol - The IP protocol name (tcp, udp, icmp, icmpv6) or number (see Protocol Numbers).

VPC only

Use -1 to specify all protocols. When authorizing security group rules, specifying -1 or a protocol number other than tcp, udp, icmp, or icmpv6 allows traffic on all ports, regardless of any port range you specify. For tcp, udp, and icmp, you must specify a port range. For icmpv6, the port range is optional; if you omit the port range, traffic for all types and codes is allowed.

ipPermission_fromPort :: Lens' IpPermission (Maybe Int)

The start of port range for the TCP and UDP protocols, or an ICMP/ICMPv6 type number. A value of -1 indicates all ICMP/ICMPv6 types. If you specify all ICMP/ICMPv6 types, you must specify all codes.

ipPermission_userIdGroupPairs :: Lens' IpPermission (Maybe [UserIdGroupPair])

The security group and Amazon Web Services account ID pairs.

ipPermission_prefixListIds :: Lens' IpPermission (Maybe [PrefixListId])

VPC only

The prefix list IDs.

ipPermission_toPort :: Lens' IpPermission (Maybe Int)

The end of port range for the TCP and UDP protocols, or an ICMP/ICMPv6 code. A value of -1 indicates all ICMP/ICMPv6 codes. If you specify all ICMP/ICMPv6 types, you must specify all codes.

ipPermission_ipv6Ranges :: Lens' IpPermission (Maybe [Ipv6Range])

VPC only

The IPv6 ranges.

ipPermission_ipRanges :: Lens' IpPermission (Maybe [IpRange])

The IPv4 ranges.

ipPermission_ipProtocol :: Lens' IpPermission Text

The IP protocol name (tcp, udp, icmp, icmpv6) or number (see Protocol Numbers).

VPC only

Use -1 to specify all protocols. When authorizing security group rules, specifying -1 or a protocol number other than tcp, udp, icmp, or icmpv6 allows traffic on all ports, regardless of any port range you specify. For tcp, udp, and icmp, you must specify a port range. For icmpv6, the port range is optional; if you omit the port range, traffic for all types and codes is allowed.