Network.AWS.EC2.Types.LaunchTemplateCapacityReservationSpecificationResponse

Description

Documentation

data LaunchTemplateCapacityReservationSpecificationResponse

Information about the Capacity Reservation targeting option.

See: newLaunchTemplateCapacityReservationSpecificationResponse smart constructor.

Constructors

LaunchTemplateCapacityReservationSpecificationResponse'

Fields

capacityReservationPreference :: Maybe CapacityReservationPreference

Indicates the instance's Capacity Reservation preferences. Possible preferences include:

open - The instance can run in any open Capacity Reservation that has matching attributes (instance type, platform, Availability Zone).

none - The instance avoids running in a Capacity Reservation even if one is available. The instance runs in On-Demand capacity.

capacityReservationTarget :: Maybe CapacityReservationTargetResponse

Information about the target Capacity Reservation or Capacity Reservation group.

newLaunchTemplateCapacityReservationSpecificationResponse :: LaunchTemplateCapacityReservationSpecificationResponse

Create a value of LaunchTemplateCapacityReservationSpecificationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:capacityReservationPreference:LaunchTemplateCapacityReservationSpecificationResponse', launchTemplateCapacityReservationSpecificationResponse_capacityReservationPreference - Indicates the instance's Capacity Reservation preferences. Possible preferences include:

open - The instance can run in any open Capacity Reservation that has matching attributes (instance type, platform, Availability Zone).

none - The instance avoids running in a Capacity Reservation even if one is available. The instance runs in On-Demand capacity.

$sel:capacityReservationTarget:LaunchTemplateCapacityReservationSpecificationResponse', launchTemplateCapacityReservationSpecificationResponse_capacityReservationTarget - Information about the target Capacity Reservation or Capacity Reservation group.

launchTemplateCapacityReservationSpecificationResponse_capacityReservationPreference :: Lens' LaunchTemplateCapacityReservationSpecificationResponse (Maybe CapacityReservationPreference)

Indicates the instance's Capacity Reservation preferences. Possible preferences include:

open - The instance can run in any open Capacity Reservation that has matching attributes (instance type, platform, Availability Zone).

none - The instance avoids running in a Capacity Reservation even if one is available. The instance runs in On-Demand capacity.

launchTemplateCapacityReservationSpecificationResponse_capacityReservationTarget :: Lens' LaunchTemplateCapacityReservationSpecificationResponse (Maybe CapacityReservationTargetResponse)

Information about the target Capacity Reservation or Capacity Reservation group.