Network.AWS.EC2.Types.LaunchTemplateInstanceMetadataOptions

Description

Documentation

data LaunchTemplateInstanceMetadataOptions

The metadata options for the instance. For more information, see Instance Metadata and User Data in the Amazon Elastic Compute Cloud User Guide.

See: newLaunchTemplateInstanceMetadataOptions smart constructor.

Constructors

LaunchTemplateInstanceMetadataOptions'

Fields

httpProtocolIpv6 :: Maybe LaunchTemplateInstanceMetadataProtocolIpv6

Enables or disables the IPv6 endpoint for the instance metadata service.

Default: disabled

httpEndpoint :: Maybe LaunchTemplateInstanceMetadataEndpointState

This parameter enables or disables the HTTP metadata endpoint on your instances. If the parameter is not specified, the default state is enabled.

If you specify a value of disabled, you will not be able to access your instance metadata.

httpPutResponseHopLimit :: Maybe Int

The desired HTTP PUT response hop limit for instance metadata requests. The larger the number, the further instance metadata requests can travel.

Default: 1

Possible values: Integers from 1 to 64

state :: Maybe LaunchTemplateInstanceMetadataOptionsState

The state of the metadata option changes.

pending - The metadata options are being updated and the instance is not ready to process metadata traffic with the new selection.

applied - The metadata options have been successfully applied on the instance.

httpTokens :: Maybe LaunchTemplateHttpTokensState

The state of token usage for your instance metadata requests. If the parameter is not specified in the request, the default state is optional.

If the state is optional, you can choose to retrieve instance metadata with or without a signed token header on your request. If you retrieve the IAM role credentials without a token, the version 1.0 role credentials are returned. If you retrieve the IAM role credentials using a valid signed token, the version 2.0 role credentials are returned.

If the state is required, you must send a signed token header with any instance metadata retrieval requests. In this state, retrieving the IAM role credentials always returns the version 2.0 credentials; the version 1.0 credentials are not available.

newLaunchTemplateInstanceMetadataOptions :: LaunchTemplateInstanceMetadataOptions

Create a value of LaunchTemplateInstanceMetadataOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpProtocolIpv6:LaunchTemplateInstanceMetadataOptions', launchTemplateInstanceMetadataOptions_httpProtocolIpv6 - Enables or disables the IPv6 endpoint for the instance metadata service.

Default: disabled

$sel:httpEndpoint:LaunchTemplateInstanceMetadataOptions', launchTemplateInstanceMetadataOptions_httpEndpoint - This parameter enables or disables the HTTP metadata endpoint on your instances. If the parameter is not specified, the default state is enabled.

If you specify a value of disabled, you will not be able to access your instance metadata.

$sel:httpPutResponseHopLimit:LaunchTemplateInstanceMetadataOptions', launchTemplateInstanceMetadataOptions_httpPutResponseHopLimit - The desired HTTP PUT response hop limit for instance metadata requests. The larger the number, the further instance metadata requests can travel.

Default: 1

Possible values: Integers from 1 to 64

$sel:state:LaunchTemplateInstanceMetadataOptions', launchTemplateInstanceMetadataOptions_state - The state of the metadata option changes.

pending - The metadata options are being updated and the instance is not ready to process metadata traffic with the new selection.

applied - The metadata options have been successfully applied on the instance.

$sel:httpTokens:LaunchTemplateInstanceMetadataOptions', launchTemplateInstanceMetadataOptions_httpTokens - The state of token usage for your instance metadata requests. If the parameter is not specified in the request, the default state is optional.

If the state is optional, you can choose to retrieve instance metadata with or without a signed token header on your request. If you retrieve the IAM role credentials without a token, the version 1.0 role credentials are returned. If you retrieve the IAM role credentials using a valid signed token, the version 2.0 role credentials are returned.

If the state is required, you must send a signed token header with any instance metadata retrieval requests. In this state, retrieving the IAM role credentials always returns the version 2.0 credentials; the version 1.0 credentials are not available.

launchTemplateInstanceMetadataOptions_httpProtocolIpv6 :: Lens' LaunchTemplateInstanceMetadataOptions (Maybe LaunchTemplateInstanceMetadataProtocolIpv6)

Enables or disables the IPv6 endpoint for the instance metadata service.

Default: disabled

launchTemplateInstanceMetadataOptions_httpEndpoint :: Lens' LaunchTemplateInstanceMetadataOptions (Maybe LaunchTemplateInstanceMetadataEndpointState)

This parameter enables or disables the HTTP metadata endpoint on your instances. If the parameter is not specified, the default state is enabled.

If you specify a value of disabled, you will not be able to access your instance metadata.

launchTemplateInstanceMetadataOptions_httpPutResponseHopLimit :: Lens' LaunchTemplateInstanceMetadataOptions (Maybe Int)

The desired HTTP PUT response hop limit for instance metadata requests. The larger the number, the further instance metadata requests can travel.

Default: 1

Possible values: Integers from 1 to 64

launchTemplateInstanceMetadataOptions_state :: Lens' LaunchTemplateInstanceMetadataOptions (Maybe LaunchTemplateInstanceMetadataOptionsState)

The state of the metadata option changes.

pending - The metadata options are being updated and the instance is not ready to process metadata traffic with the new selection.

applied - The metadata options have been successfully applied on the instance.

launchTemplateInstanceMetadataOptions_httpTokens :: Lens' LaunchTemplateInstanceMetadataOptions (Maybe LaunchTemplateHttpTokensState)

The state of token usage for your instance metadata requests. If the parameter is not specified in the request, the default state is optional.

If the state is optional, you can choose to retrieve instance metadata with or without a signed token header on your request. If you retrieve the IAM role credentials without a token, the version 1.0 role credentials are returned. If you retrieve the IAM role credentials using a valid signed token, the version 2.0 role credentials are returned.

If the state is required, you must send a signed token header with any instance metadata retrieval requests. In this state, retrieving the IAM role credentials always returns the version 2.0 credentials; the version 1.0 credentials are not available.