Network.AWS.EC2.Types.LaunchTemplateInstanceMetadataOptionsRequest

Description

Documentation

data LaunchTemplateInstanceMetadataOptionsRequest

The metadata options for the instance. For more information, see Instance Metadata and User Data in the Amazon Elastic Compute Cloud User Guide.

See: newLaunchTemplateInstanceMetadataOptionsRequest smart constructor.

Constructors

LaunchTemplateInstanceMetadataOptionsRequest'

Fields

httpProtocolIpv6 :: Maybe LaunchTemplateInstanceMetadataProtocolIpv6

Enables or disables the IPv6 endpoint for the instance metadata service.

Default: disabled

httpEndpoint :: Maybe LaunchTemplateInstanceMetadataEndpointState

This parameter enables or disables the HTTP metadata endpoint on your instances. If the parameter is not specified, the default state is enabled.

If you specify a value of disabled, you will not be able to access your instance metadata.

httpPutResponseHopLimit :: Maybe Int

The desired HTTP PUT response hop limit for instance metadata requests. The larger the number, the further instance metadata requests can travel.

Default: 1

Possible values: Integers from 1 to 64

httpTokens :: Maybe LaunchTemplateHttpTokensState

The state of token usage for your instance metadata requests. If the parameter is not specified in the request, the default state is optional.

If the state is optional, you can choose to retrieve instance metadata with or without a signed token header on your request. If you retrieve the IAM role credentials without a token, the version 1.0 role credentials are returned. If you retrieve the IAM role credentials using a valid signed token, the version 2.0 role credentials are returned.

If the state is required, you must send a signed token header with any instance metadata retrieval requests. In this state, retrieving the IAM role credentials always returns the version 2.0 credentials; the version 1.0 credentials are not available.

newLaunchTemplateInstanceMetadataOptionsRequest :: LaunchTemplateInstanceMetadataOptionsRequest

Create a value of LaunchTemplateInstanceMetadataOptionsRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpProtocolIpv6:LaunchTemplateInstanceMetadataOptionsRequest', launchTemplateInstanceMetadataOptionsRequest_httpProtocolIpv6 - Enables or disables the IPv6 endpoint for the instance metadata service.

Default: disabled

$sel:httpEndpoint:LaunchTemplateInstanceMetadataOptionsRequest', launchTemplateInstanceMetadataOptionsRequest_httpEndpoint - This parameter enables or disables the HTTP metadata endpoint on your instances. If the parameter is not specified, the default state is enabled.

If you specify a value of disabled, you will not be able to access your instance metadata.

$sel:httpPutResponseHopLimit:LaunchTemplateInstanceMetadataOptionsRequest', launchTemplateInstanceMetadataOptionsRequest_httpPutResponseHopLimit - The desired HTTP PUT response hop limit for instance metadata requests. The larger the number, the further instance metadata requests can travel.

Default: 1

Possible values: Integers from 1 to 64

$sel:httpTokens:LaunchTemplateInstanceMetadataOptionsRequest', launchTemplateInstanceMetadataOptionsRequest_httpTokens - The state of token usage for your instance metadata requests. If the parameter is not specified in the request, the default state is optional.

If the state is optional, you can choose to retrieve instance metadata with or without a signed token header on your request. If you retrieve the IAM role credentials without a token, the version 1.0 role credentials are returned. If you retrieve the IAM role credentials using a valid signed token, the version 2.0 role credentials are returned.

If the state is required, you must send a signed token header with any instance metadata retrieval requests. In this state, retrieving the IAM role credentials always returns the version 2.0 credentials; the version 1.0 credentials are not available.

launchTemplateInstanceMetadataOptionsRequest_httpProtocolIpv6 :: Lens' LaunchTemplateInstanceMetadataOptionsRequest (Maybe LaunchTemplateInstanceMetadataProtocolIpv6)

Enables or disables the IPv6 endpoint for the instance metadata service.

Default: disabled

launchTemplateInstanceMetadataOptionsRequest_httpEndpoint :: Lens' LaunchTemplateInstanceMetadataOptionsRequest (Maybe LaunchTemplateInstanceMetadataEndpointState)

This parameter enables or disables the HTTP metadata endpoint on your instances. If the parameter is not specified, the default state is enabled.

If you specify a value of disabled, you will not be able to access your instance metadata.

launchTemplateInstanceMetadataOptionsRequest_httpPutResponseHopLimit :: Lens' LaunchTemplateInstanceMetadataOptionsRequest (Maybe Int)

The desired HTTP PUT response hop limit for instance metadata requests. The larger the number, the further instance metadata requests can travel.

Default: 1

Possible values: Integers from 1 to 64

launchTemplateInstanceMetadataOptionsRequest_httpTokens :: Lens' LaunchTemplateInstanceMetadataOptionsRequest (Maybe LaunchTemplateHttpTokensState)

The state of token usage for your instance metadata requests. If the parameter is not specified in the request, the default state is optional.

If the state is optional, you can choose to retrieve instance metadata with or without a signed token header on your request. If you retrieve the IAM role credentials without a token, the version 1.0 role credentials are returned. If you retrieve the IAM role credentials using a valid signed token, the version 2.0 role credentials are returned.

If the state is required, you must send a signed token header with any instance metadata retrieval requests. In this state, retrieving the IAM role credentials always returns the version 2.0 credentials; the version 1.0 credentials are not available.