Network.AWS.EC2.Types.LaunchTemplateInstanceNetworkInterfaceSpecification

Description

Documentation

data LaunchTemplateInstanceNetworkInterfaceSpecification

Describes a network interface.

See: newLaunchTemplateInstanceNetworkInterfaceSpecification smart constructor.

Constructors

LaunchTemplateInstanceNetworkInterfaceSpecification'

Fields

groups :: Maybe [Text]

The IDs of one or more security groups.

privateIpAddresses :: Maybe [PrivateIpAddressSpecification]

One or more private IPv4 addresses.

deleteOnTermination :: Maybe Bool

Indicates whether the network interface is deleted when the instance is terminated.

associateCarrierIpAddress :: Maybe Bool

Indicates whether to associate a Carrier IP address with eth0 for a new network interface.

Use this option when you launch an instance in a Wavelength Zone and want to associate a Carrier IP address with the network interface. For more information about Carrier IP addresses, see Carrier IP addresses in the Wavelength Developer Guide.

associatePublicIpAddress :: Maybe Bool

Indicates whether to associate a public IPv4 address with eth0 for a new network interface.

ipv4Prefixes :: Maybe [Ipv4PrefixSpecificationResponse]

One or more IPv4 prefixes assigned to the network interface.

interfaceType :: Maybe Text

The type of network interface.

ipv4PrefixCount :: Maybe Int

The number of IPv4 prefixes that Amazon Web Services automatically assigned to the network interface.

networkInterfaceId :: Maybe Text

The ID of the network interface.

subnetId :: Maybe Text

The ID of the subnet for the network interface.

ipv6AddressCount :: Maybe Int

The number of IPv6 addresses for the network interface.

networkCardIndex :: Maybe Int

The index of the network card.

ipv6Prefixes :: Maybe [Ipv6PrefixSpecificationResponse]

One or more IPv6 prefixes assigned to the network interface.

privateIpAddress :: Maybe Text

The primary private IPv4 address of the network interface.

ipv6PrefixCount :: Maybe Int

The number of IPv6 prefixes that Amazon Web Services automatically assigned to the network interface.

secondaryPrivateIpAddressCount :: Maybe Int

The number of secondary private IPv4 addresses for the network interface.

description :: Maybe Text

A description for the network interface.

deviceIndex :: Maybe Int

The device index for the network interface attachment.

ipv6Addresses :: Maybe [InstanceIpv6Address]

The IPv6 addresses for the network interface.

newLaunchTemplateInstanceNetworkInterfaceSpecification :: LaunchTemplateInstanceNetworkInterfaceSpecification

Create a value of LaunchTemplateInstanceNetworkInterfaceSpecification with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groups:LaunchTemplateInstanceNetworkInterfaceSpecification', launchTemplateInstanceNetworkInterfaceSpecification_groups - The IDs of one or more security groups.

$sel:privateIpAddresses:LaunchTemplateInstanceNetworkInterfaceSpecification', launchTemplateInstanceNetworkInterfaceSpecification_privateIpAddresses - One or more private IPv4 addresses.

$sel:deleteOnTermination:LaunchTemplateInstanceNetworkInterfaceSpecification', launchTemplateInstanceNetworkInterfaceSpecification_deleteOnTermination - Indicates whether the network interface is deleted when the instance is terminated.

$sel:associateCarrierIpAddress:LaunchTemplateInstanceNetworkInterfaceSpecification', launchTemplateInstanceNetworkInterfaceSpecification_associateCarrierIpAddress - Indicates whether to associate a Carrier IP address with eth0 for a new network interface.

Use this option when you launch an instance in a Wavelength Zone and want to associate a Carrier IP address with the network interface. For more information about Carrier IP addresses, see Carrier IP addresses in the Wavelength Developer Guide.

$sel:associatePublicIpAddress:LaunchTemplateInstanceNetworkInterfaceSpecification', launchTemplateInstanceNetworkInterfaceSpecification_associatePublicIpAddress - Indicates whether to associate a public IPv4 address with eth0 for a new network interface.

$sel:ipv4Prefixes:LaunchTemplateInstanceNetworkInterfaceSpecification', launchTemplateInstanceNetworkInterfaceSpecification_ipv4Prefixes - One or more IPv4 prefixes assigned to the network interface.

$sel:interfaceType:LaunchTemplateInstanceNetworkInterfaceSpecification', launchTemplateInstanceNetworkInterfaceSpecification_interfaceType - The type of network interface.

$sel:ipv4PrefixCount:LaunchTemplateInstanceNetworkInterfaceSpecification', launchTemplateInstanceNetworkInterfaceSpecification_ipv4PrefixCount - The number of IPv4 prefixes that Amazon Web Services automatically assigned to the network interface.

$sel:networkInterfaceId:LaunchTemplateInstanceNetworkInterfaceSpecification', launchTemplateInstanceNetworkInterfaceSpecification_networkInterfaceId - The ID of the network interface.

$sel:subnetId:LaunchTemplateInstanceNetworkInterfaceSpecification', launchTemplateInstanceNetworkInterfaceSpecification_subnetId - The ID of the subnet for the network interface.

$sel:ipv6AddressCount:LaunchTemplateInstanceNetworkInterfaceSpecification', launchTemplateInstanceNetworkInterfaceSpecification_ipv6AddressCount - The number of IPv6 addresses for the network interface.

$sel:networkCardIndex:LaunchTemplateInstanceNetworkInterfaceSpecification', launchTemplateInstanceNetworkInterfaceSpecification_networkCardIndex - The index of the network card.

$sel:ipv6Prefixes:LaunchTemplateInstanceNetworkInterfaceSpecification', launchTemplateInstanceNetworkInterfaceSpecification_ipv6Prefixes - One or more IPv6 prefixes assigned to the network interface.

$sel:privateIpAddress:LaunchTemplateInstanceNetworkInterfaceSpecification', launchTemplateInstanceNetworkInterfaceSpecification_privateIpAddress - The primary private IPv4 address of the network interface.

$sel:ipv6PrefixCount:LaunchTemplateInstanceNetworkInterfaceSpecification', launchTemplateInstanceNetworkInterfaceSpecification_ipv6PrefixCount - The number of IPv6 prefixes that Amazon Web Services automatically assigned to the network interface.

$sel:secondaryPrivateIpAddressCount:LaunchTemplateInstanceNetworkInterfaceSpecification', launchTemplateInstanceNetworkInterfaceSpecification_secondaryPrivateIpAddressCount - The number of secondary private IPv4 addresses for the network interface.

$sel:description:LaunchTemplateInstanceNetworkInterfaceSpecification', launchTemplateInstanceNetworkInterfaceSpecification_description - A description for the network interface.

$sel:deviceIndex:LaunchTemplateInstanceNetworkInterfaceSpecification', launchTemplateInstanceNetworkInterfaceSpecification_deviceIndex - The device index for the network interface attachment.

$sel:ipv6Addresses:LaunchTemplateInstanceNetworkInterfaceSpecification', launchTemplateInstanceNetworkInterfaceSpecification_ipv6Addresses - The IPv6 addresses for the network interface.

launchTemplateInstanceNetworkInterfaceSpecification_groups :: Lens' LaunchTemplateInstanceNetworkInterfaceSpecification (Maybe [Text])

The IDs of one or more security groups.

launchTemplateInstanceNetworkInterfaceSpecification_privateIpAddresses :: Lens' LaunchTemplateInstanceNetworkInterfaceSpecification (Maybe [PrivateIpAddressSpecification])

One or more private IPv4 addresses.

launchTemplateInstanceNetworkInterfaceSpecification_deleteOnTermination :: Lens' LaunchTemplateInstanceNetworkInterfaceSpecification (Maybe Bool)

Indicates whether the network interface is deleted when the instance is terminated.

launchTemplateInstanceNetworkInterfaceSpecification_associateCarrierIpAddress :: Lens' LaunchTemplateInstanceNetworkInterfaceSpecification (Maybe Bool)

Indicates whether to associate a Carrier IP address with eth0 for a new network interface.

Use this option when you launch an instance in a Wavelength Zone and want to associate a Carrier IP address with the network interface. For more information about Carrier IP addresses, see Carrier IP addresses in the Wavelength Developer Guide.

launchTemplateInstanceNetworkInterfaceSpecification_associatePublicIpAddress :: Lens' LaunchTemplateInstanceNetworkInterfaceSpecification (Maybe Bool)

Indicates whether to associate a public IPv4 address with eth0 for a new network interface.

launchTemplateInstanceNetworkInterfaceSpecification_ipv4Prefixes :: Lens' LaunchTemplateInstanceNetworkInterfaceSpecification (Maybe [Ipv4PrefixSpecificationResponse])

One or more IPv4 prefixes assigned to the network interface.

launchTemplateInstanceNetworkInterfaceSpecification_interfaceType :: Lens' LaunchTemplateInstanceNetworkInterfaceSpecification (Maybe Text)

The type of network interface.

launchTemplateInstanceNetworkInterfaceSpecification_ipv4PrefixCount :: Lens' LaunchTemplateInstanceNetworkInterfaceSpecification (Maybe Int)

The number of IPv4 prefixes that Amazon Web Services automatically assigned to the network interface.

launchTemplateInstanceNetworkInterfaceSpecification_networkInterfaceId :: Lens' LaunchTemplateInstanceNetworkInterfaceSpecification (Maybe Text)

The ID of the network interface.

launchTemplateInstanceNetworkInterfaceSpecification_subnetId :: Lens' LaunchTemplateInstanceNetworkInterfaceSpecification (Maybe Text)

The ID of the subnet for the network interface.

launchTemplateInstanceNetworkInterfaceSpecification_ipv6AddressCount :: Lens' LaunchTemplateInstanceNetworkInterfaceSpecification (Maybe Int)

The number of IPv6 addresses for the network interface.

launchTemplateInstanceNetworkInterfaceSpecification_networkCardIndex :: Lens' LaunchTemplateInstanceNetworkInterfaceSpecification (Maybe Int)

The index of the network card.

launchTemplateInstanceNetworkInterfaceSpecification_ipv6Prefixes :: Lens' LaunchTemplateInstanceNetworkInterfaceSpecification (Maybe [Ipv6PrefixSpecificationResponse])

One or more IPv6 prefixes assigned to the network interface.

launchTemplateInstanceNetworkInterfaceSpecification_privateIpAddress :: Lens' LaunchTemplateInstanceNetworkInterfaceSpecification (Maybe Text)

The primary private IPv4 address of the network interface.

launchTemplateInstanceNetworkInterfaceSpecification_ipv6PrefixCount :: Lens' LaunchTemplateInstanceNetworkInterfaceSpecification (Maybe Int)

The number of IPv6 prefixes that Amazon Web Services automatically assigned to the network interface.

launchTemplateInstanceNetworkInterfaceSpecification_secondaryPrivateIpAddressCount :: Lens' LaunchTemplateInstanceNetworkInterfaceSpecification (Maybe Int)

The number of secondary private IPv4 addresses for the network interface.

launchTemplateInstanceNetworkInterfaceSpecification_description :: Lens' LaunchTemplateInstanceNetworkInterfaceSpecification (Maybe Text)

A description for the network interface.

launchTemplateInstanceNetworkInterfaceSpecification_deviceIndex :: Lens' LaunchTemplateInstanceNetworkInterfaceSpecification (Maybe Int)

The device index for the network interface attachment.

launchTemplateInstanceNetworkInterfaceSpecification_ipv6Addresses :: Lens' LaunchTemplateInstanceNetworkInterfaceSpecification (Maybe [InstanceIpv6Address])

The IPv6 addresses for the network interface.