Network.AWS.EC2.Types.LaunchTemplateInstanceNetworkInterfaceSpecificationRequest

Description

Documentation

data LaunchTemplateInstanceNetworkInterfaceSpecificationRequest

The parameters for a network interface.

See: newLaunchTemplateInstanceNetworkInterfaceSpecificationRequest smart constructor.

Constructors

LaunchTemplateInstanceNetworkInterfaceSpecificationRequest'

Fields

groups :: Maybe [Text]

The IDs of one or more security groups.

privateIpAddresses :: Maybe [PrivateIpAddressSpecification]

One or more private IPv4 addresses.

deleteOnTermination :: Maybe Bool

Indicates whether the network interface is deleted when the instance is terminated.

associateCarrierIpAddress :: Maybe Bool

Associates a Carrier IP address with eth0 for a new network interface.

Use this option when you launch an instance in a Wavelength Zone and want to associate a Carrier IP address with the network interface. For more information about Carrier IP addresses, see Carrier IP addresses in the Wavelength Developer Guide.

associatePublicIpAddress :: Maybe Bool

Associates a public IPv4 address with eth0 for a new network interface.

ipv4Prefixes :: Maybe [Ipv4PrefixSpecificationRequest]

One or more IPv4 prefixes to be assigned to the network interface. You cannot use this option if you use the Ipv4PrefixCount option.

interfaceType :: Maybe Text

The type of network interface. To create an Elastic Fabric Adapter (EFA), specify efa. For more information, see Elastic Fabric Adapter in the Amazon Elastic Compute Cloud User Guide.

If you are not creating an EFA, specify interface or omit this parameter.

Valid values: interface | efa

ipv4PrefixCount :: Maybe Int

The number of IPv4 prefixes to be automatically assigned to the network interface. You cannot use this option if you use the Ipv4Prefix option.

networkInterfaceId :: Maybe Text

The ID of the network interface.

subnetId :: Maybe Text

The ID of the subnet for the network interface.

ipv6AddressCount :: Maybe Int

The number of IPv6 addresses to assign to a network interface. Amazon EC2 automatically selects the IPv6 addresses from the subnet range. You can't use this option if specifying specific IPv6 addresses.

networkCardIndex :: Maybe Int

The index of the network card. Some instance types support multiple network cards. The primary network interface must be assigned to network card index 0. The default is network card index 0.

ipv6Prefixes :: Maybe [Ipv6PrefixSpecificationRequest]

One or more IPv6 prefixes to be assigned to the network interface. You cannot use this option if you use the Ipv6PrefixCount option.

privateIpAddress :: Maybe Text

The primary private IPv4 address of the network interface.

ipv6PrefixCount :: Maybe Int

The number of IPv6 prefixes to be automatically assigned to the network interface. You cannot use this option if you use the Ipv6Prefix option.

secondaryPrivateIpAddressCount :: Maybe Int

The number of secondary private IPv4 addresses to assign to a network interface.

description :: Maybe Text

A description for the network interface.

deviceIndex :: Maybe Int

The device index for the network interface attachment.

ipv6Addresses :: Maybe [InstanceIpv6AddressRequest]

One or more specific IPv6 addresses from the IPv6 CIDR block range of your subnet. You can't use this option if you're specifying a number of IPv6 addresses.

newLaunchTemplateInstanceNetworkInterfaceSpecificationRequest :: LaunchTemplateInstanceNetworkInterfaceSpecificationRequest

Create a value of LaunchTemplateInstanceNetworkInterfaceSpecificationRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groups:LaunchTemplateInstanceNetworkInterfaceSpecificationRequest', launchTemplateInstanceNetworkInterfaceSpecificationRequest_groups - The IDs of one or more security groups.

$sel:privateIpAddresses:LaunchTemplateInstanceNetworkInterfaceSpecificationRequest', launchTemplateInstanceNetworkInterfaceSpecificationRequest_privateIpAddresses - One or more private IPv4 addresses.

$sel:deleteOnTermination:LaunchTemplateInstanceNetworkInterfaceSpecificationRequest', launchTemplateInstanceNetworkInterfaceSpecificationRequest_deleteOnTermination - Indicates whether the network interface is deleted when the instance is terminated.

$sel:associateCarrierIpAddress:LaunchTemplateInstanceNetworkInterfaceSpecificationRequest', launchTemplateInstanceNetworkInterfaceSpecificationRequest_associateCarrierIpAddress - Associates a Carrier IP address with eth0 for a new network interface.

Use this option when you launch an instance in a Wavelength Zone and want to associate a Carrier IP address with the network interface. For more information about Carrier IP addresses, see Carrier IP addresses in the Wavelength Developer Guide.

$sel:associatePublicIpAddress:LaunchTemplateInstanceNetworkInterfaceSpecificationRequest', launchTemplateInstanceNetworkInterfaceSpecificationRequest_associatePublicIpAddress - Associates a public IPv4 address with eth0 for a new network interface.

$sel:ipv4Prefixes:LaunchTemplateInstanceNetworkInterfaceSpecificationRequest', launchTemplateInstanceNetworkInterfaceSpecificationRequest_ipv4Prefixes - One or more IPv4 prefixes to be assigned to the network interface. You cannot use this option if you use the Ipv4PrefixCount option.

$sel:interfaceType:LaunchTemplateInstanceNetworkInterfaceSpecificationRequest', launchTemplateInstanceNetworkInterfaceSpecificationRequest_interfaceType - The type of network interface. To create an Elastic Fabric Adapter (EFA), specify efa. For more information, see Elastic Fabric Adapter in the Amazon Elastic Compute Cloud User Guide.

If you are not creating an EFA, specify interface or omit this parameter.

Valid values: interface | efa

$sel:ipv4PrefixCount:LaunchTemplateInstanceNetworkInterfaceSpecificationRequest', launchTemplateInstanceNetworkInterfaceSpecificationRequest_ipv4PrefixCount - The number of IPv4 prefixes to be automatically assigned to the network interface. You cannot use this option if you use the Ipv4Prefix option.

$sel:networkInterfaceId:LaunchTemplateInstanceNetworkInterfaceSpecificationRequest', launchTemplateInstanceNetworkInterfaceSpecificationRequest_networkInterfaceId - The ID of the network interface.

$sel:subnetId:LaunchTemplateInstanceNetworkInterfaceSpecificationRequest', launchTemplateInstanceNetworkInterfaceSpecificationRequest_subnetId - The ID of the subnet for the network interface.

$sel:ipv6AddressCount:LaunchTemplateInstanceNetworkInterfaceSpecificationRequest', launchTemplateInstanceNetworkInterfaceSpecificationRequest_ipv6AddressCount - The number of IPv6 addresses to assign to a network interface. Amazon EC2 automatically selects the IPv6 addresses from the subnet range. You can't use this option if specifying specific IPv6 addresses.

$sel:networkCardIndex:LaunchTemplateInstanceNetworkInterfaceSpecificationRequest', launchTemplateInstanceNetworkInterfaceSpecificationRequest_networkCardIndex - The index of the network card. Some instance types support multiple network cards. The primary network interface must be assigned to network card index 0. The default is network card index 0.

$sel:ipv6Prefixes:LaunchTemplateInstanceNetworkInterfaceSpecificationRequest', launchTemplateInstanceNetworkInterfaceSpecificationRequest_ipv6Prefixes - One or more IPv6 prefixes to be assigned to the network interface. You cannot use this option if you use the Ipv6PrefixCount option.

$sel:privateIpAddress:LaunchTemplateInstanceNetworkInterfaceSpecificationRequest', launchTemplateInstanceNetworkInterfaceSpecificationRequest_privateIpAddress - The primary private IPv4 address of the network interface.

$sel:ipv6PrefixCount:LaunchTemplateInstanceNetworkInterfaceSpecificationRequest', launchTemplateInstanceNetworkInterfaceSpecificationRequest_ipv6PrefixCount - The number of IPv6 prefixes to be automatically assigned to the network interface. You cannot use this option if you use the Ipv6Prefix option.

$sel:secondaryPrivateIpAddressCount:LaunchTemplateInstanceNetworkInterfaceSpecificationRequest', launchTemplateInstanceNetworkInterfaceSpecificationRequest_secondaryPrivateIpAddressCount - The number of secondary private IPv4 addresses to assign to a network interface.

$sel:description:LaunchTemplateInstanceNetworkInterfaceSpecificationRequest', launchTemplateInstanceNetworkInterfaceSpecificationRequest_description - A description for the network interface.

$sel:deviceIndex:LaunchTemplateInstanceNetworkInterfaceSpecificationRequest', launchTemplateInstanceNetworkInterfaceSpecificationRequest_deviceIndex - The device index for the network interface attachment.

$sel:ipv6Addresses:LaunchTemplateInstanceNetworkInterfaceSpecificationRequest', launchTemplateInstanceNetworkInterfaceSpecificationRequest_ipv6Addresses - One or more specific IPv6 addresses from the IPv6 CIDR block range of your subnet. You can't use this option if you're specifying a number of IPv6 addresses.

launchTemplateInstanceNetworkInterfaceSpecificationRequest_groups :: Lens' LaunchTemplateInstanceNetworkInterfaceSpecificationRequest (Maybe [Text])

The IDs of one or more security groups.

launchTemplateInstanceNetworkInterfaceSpecificationRequest_privateIpAddresses :: Lens' LaunchTemplateInstanceNetworkInterfaceSpecificationRequest (Maybe [PrivateIpAddressSpecification])

One or more private IPv4 addresses.

launchTemplateInstanceNetworkInterfaceSpecificationRequest_deleteOnTermination :: Lens' LaunchTemplateInstanceNetworkInterfaceSpecificationRequest (Maybe Bool)

Indicates whether the network interface is deleted when the instance is terminated.

launchTemplateInstanceNetworkInterfaceSpecificationRequest_associateCarrierIpAddress :: Lens' LaunchTemplateInstanceNetworkInterfaceSpecificationRequest (Maybe Bool)

Associates a Carrier IP address with eth0 for a new network interface.

Use this option when you launch an instance in a Wavelength Zone and want to associate a Carrier IP address with the network interface. For more information about Carrier IP addresses, see Carrier IP addresses in the Wavelength Developer Guide.

launchTemplateInstanceNetworkInterfaceSpecificationRequest_associatePublicIpAddress :: Lens' LaunchTemplateInstanceNetworkInterfaceSpecificationRequest (Maybe Bool)

Associates a public IPv4 address with eth0 for a new network interface.

launchTemplateInstanceNetworkInterfaceSpecificationRequest_ipv4Prefixes :: Lens' LaunchTemplateInstanceNetworkInterfaceSpecificationRequest (Maybe [Ipv4PrefixSpecificationRequest])

One or more IPv4 prefixes to be assigned to the network interface. You cannot use this option if you use the Ipv4PrefixCount option.

launchTemplateInstanceNetworkInterfaceSpecificationRequest_interfaceType :: Lens' LaunchTemplateInstanceNetworkInterfaceSpecificationRequest (Maybe Text)

The type of network interface. To create an Elastic Fabric Adapter (EFA), specify efa. For more information, see Elastic Fabric Adapter in the Amazon Elastic Compute Cloud User Guide.

If you are not creating an EFA, specify interface or omit this parameter.

Valid values: interface | efa

launchTemplateInstanceNetworkInterfaceSpecificationRequest_ipv4PrefixCount :: Lens' LaunchTemplateInstanceNetworkInterfaceSpecificationRequest (Maybe Int)

The number of IPv4 prefixes to be automatically assigned to the network interface. You cannot use this option if you use the Ipv4Prefix option.

launchTemplateInstanceNetworkInterfaceSpecificationRequest_networkInterfaceId :: Lens' LaunchTemplateInstanceNetworkInterfaceSpecificationRequest (Maybe Text)

The ID of the network interface.

launchTemplateInstanceNetworkInterfaceSpecificationRequest_subnetId :: Lens' LaunchTemplateInstanceNetworkInterfaceSpecificationRequest (Maybe Text)

The ID of the subnet for the network interface.

launchTemplateInstanceNetworkInterfaceSpecificationRequest_ipv6AddressCount :: Lens' LaunchTemplateInstanceNetworkInterfaceSpecificationRequest (Maybe Int)

The number of IPv6 addresses to assign to a network interface. Amazon EC2 automatically selects the IPv6 addresses from the subnet range. You can't use this option if specifying specific IPv6 addresses.

launchTemplateInstanceNetworkInterfaceSpecificationRequest_networkCardIndex :: Lens' LaunchTemplateInstanceNetworkInterfaceSpecificationRequest (Maybe Int)

The index of the network card. Some instance types support multiple network cards. The primary network interface must be assigned to network card index 0. The default is network card index 0.

launchTemplateInstanceNetworkInterfaceSpecificationRequest_ipv6Prefixes :: Lens' LaunchTemplateInstanceNetworkInterfaceSpecificationRequest (Maybe [Ipv6PrefixSpecificationRequest])

One or more IPv6 prefixes to be assigned to the network interface. You cannot use this option if you use the Ipv6PrefixCount option.

launchTemplateInstanceNetworkInterfaceSpecificationRequest_privateIpAddress :: Lens' LaunchTemplateInstanceNetworkInterfaceSpecificationRequest (Maybe Text)

The primary private IPv4 address of the network interface.

launchTemplateInstanceNetworkInterfaceSpecificationRequest_ipv6PrefixCount :: Lens' LaunchTemplateInstanceNetworkInterfaceSpecificationRequest (Maybe Int)

The number of IPv6 prefixes to be automatically assigned to the network interface. You cannot use this option if you use the Ipv6Prefix option.

launchTemplateInstanceNetworkInterfaceSpecificationRequest_secondaryPrivateIpAddressCount :: Lens' LaunchTemplateInstanceNetworkInterfaceSpecificationRequest (Maybe Int)

The number of secondary private IPv4 addresses to assign to a network interface.

launchTemplateInstanceNetworkInterfaceSpecificationRequest_description :: Lens' LaunchTemplateInstanceNetworkInterfaceSpecificationRequest (Maybe Text)

A description for the network interface.

launchTemplateInstanceNetworkInterfaceSpecificationRequest_deviceIndex :: Lens' LaunchTemplateInstanceNetworkInterfaceSpecificationRequest (Maybe Int)

The device index for the network interface attachment.

launchTemplateInstanceNetworkInterfaceSpecificationRequest_ipv6Addresses :: Lens' LaunchTemplateInstanceNetworkInterfaceSpecificationRequest (Maybe [InstanceIpv6AddressRequest])

One or more specific IPv6 addresses from the IPv6 CIDR block range of your subnet. You can't use this option if you're specifying a number of IPv6 addresses.