Network.AWS.EC2.Types.LaunchTemplateSpotMarketOptions

Description

Documentation

data LaunchTemplateSpotMarketOptions

The options for Spot Instances.

See: newLaunchTemplateSpotMarketOptions smart constructor.

Constructors

LaunchTemplateSpotMarketOptions'

Fields

blockDurationMinutes :: Maybe Int

The required duration for the Spot Instances (also known as Spot blocks), in minutes. This value must be a multiple of 60 (60, 120, 180, 240, 300, or 360).

instanceInterruptionBehavior :: Maybe InstanceInterruptionBehavior

The behavior when a Spot Instance is interrupted.

validUntil :: Maybe ISO8601

The end date of the request. For a one-time request, the request remains active until all instances launch, the request is canceled, or this date is reached. If the request is persistent, it remains active until it is canceled or this date and time is reached.

spotInstanceType :: Maybe SpotInstanceType

The Spot Instance request type.

maxPrice :: Maybe Text

The maximum hourly price you're willing to pay for the Spot Instances.

newLaunchTemplateSpotMarketOptions :: LaunchTemplateSpotMarketOptions

Create a value of LaunchTemplateSpotMarketOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:blockDurationMinutes:LaunchTemplateSpotMarketOptions', launchTemplateSpotMarketOptions_blockDurationMinutes - The required duration for the Spot Instances (also known as Spot blocks), in minutes. This value must be a multiple of 60 (60, 120, 180, 240, 300, or 360).

$sel:instanceInterruptionBehavior:LaunchTemplateSpotMarketOptions', launchTemplateSpotMarketOptions_instanceInterruptionBehavior - The behavior when a Spot Instance is interrupted.

$sel:validUntil:LaunchTemplateSpotMarketOptions', launchTemplateSpotMarketOptions_validUntil - The end date of the request. For a one-time request, the request remains active until all instances launch, the request is canceled, or this date is reached. If the request is persistent, it remains active until it is canceled or this date and time is reached.

$sel:spotInstanceType:LaunchTemplateSpotMarketOptions', launchTemplateSpotMarketOptions_spotInstanceType - The Spot Instance request type.

$sel:maxPrice:LaunchTemplateSpotMarketOptions', launchTemplateSpotMarketOptions_maxPrice - The maximum hourly price you're willing to pay for the Spot Instances.

launchTemplateSpotMarketOptions_blockDurationMinutes :: Lens' LaunchTemplateSpotMarketOptions (Maybe Int)

The required duration for the Spot Instances (also known as Spot blocks), in minutes. This value must be a multiple of 60 (60, 120, 180, 240, 300, or 360).

launchTemplateSpotMarketOptions_instanceInterruptionBehavior :: Lens' LaunchTemplateSpotMarketOptions (Maybe InstanceInterruptionBehavior)

The behavior when a Spot Instance is interrupted.

launchTemplateSpotMarketOptions_validUntil :: Lens' LaunchTemplateSpotMarketOptions (Maybe UTCTime)

The end date of the request. For a one-time request, the request remains active until all instances launch, the request is canceled, or this date is reached. If the request is persistent, it remains active until it is canceled or this date and time is reached.

launchTemplateSpotMarketOptions_spotInstanceType :: Lens' LaunchTemplateSpotMarketOptions (Maybe SpotInstanceType)

The Spot Instance request type.

launchTemplateSpotMarketOptions_maxPrice :: Lens' LaunchTemplateSpotMarketOptions (Maybe Text)

The maximum hourly price you're willing to pay for the Spot Instances.