Network.AWS.EC2.Types.LaunchTemplateSpotMarketOptionsRequest

Description

Documentation

data LaunchTemplateSpotMarketOptionsRequest

The options for Spot Instances.

See: newLaunchTemplateSpotMarketOptionsRequest smart constructor.

Constructors

LaunchTemplateSpotMarketOptionsRequest'

Fields

blockDurationMinutes :: Maybe Int

The required duration for the Spot Instances (also known as Spot blocks), in minutes. This value must be a multiple of 60 (60, 120, 180, 240, 300, or 360).

instanceInterruptionBehavior :: Maybe InstanceInterruptionBehavior

The behavior when a Spot Instance is interrupted. The default is terminate.

validUntil :: Maybe ISO8601

The end date of the request. For a one-time request, the request remains active until all instances launch, the request is canceled, or this date is reached. If the request is persistent, it remains active until it is canceled or this date and time is reached. The default end date is 7 days from the current date.

spotInstanceType :: Maybe SpotInstanceType

The Spot Instance request type.

maxPrice :: Maybe Text

The maximum hourly price you're willing to pay for the Spot Instances.

newLaunchTemplateSpotMarketOptionsRequest :: LaunchTemplateSpotMarketOptionsRequest

Create a value of LaunchTemplateSpotMarketOptionsRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:blockDurationMinutes:LaunchTemplateSpotMarketOptionsRequest', launchTemplateSpotMarketOptionsRequest_blockDurationMinutes - The required duration for the Spot Instances (also known as Spot blocks), in minutes. This value must be a multiple of 60 (60, 120, 180, 240, 300, or 360).

$sel:instanceInterruptionBehavior:LaunchTemplateSpotMarketOptionsRequest', launchTemplateSpotMarketOptionsRequest_instanceInterruptionBehavior - The behavior when a Spot Instance is interrupted. The default is terminate.

$sel:validUntil:LaunchTemplateSpotMarketOptionsRequest', launchTemplateSpotMarketOptionsRequest_validUntil - The end date of the request. For a one-time request, the request remains active until all instances launch, the request is canceled, or this date is reached. If the request is persistent, it remains active until it is canceled or this date and time is reached. The default end date is 7 days from the current date.

$sel:spotInstanceType:LaunchTemplateSpotMarketOptionsRequest', launchTemplateSpotMarketOptionsRequest_spotInstanceType - The Spot Instance request type.

$sel:maxPrice:LaunchTemplateSpotMarketOptionsRequest', launchTemplateSpotMarketOptionsRequest_maxPrice - The maximum hourly price you're willing to pay for the Spot Instances.

launchTemplateSpotMarketOptionsRequest_blockDurationMinutes :: Lens' LaunchTemplateSpotMarketOptionsRequest (Maybe Int)

The required duration for the Spot Instances (also known as Spot blocks), in minutes. This value must be a multiple of 60 (60, 120, 180, 240, 300, or 360).

launchTemplateSpotMarketOptionsRequest_instanceInterruptionBehavior :: Lens' LaunchTemplateSpotMarketOptionsRequest (Maybe InstanceInterruptionBehavior)

The behavior when a Spot Instance is interrupted. The default is terminate.

launchTemplateSpotMarketOptionsRequest_validUntil :: Lens' LaunchTemplateSpotMarketOptionsRequest (Maybe UTCTime)

The end date of the request. For a one-time request, the request remains active until all instances launch, the request is canceled, or this date is reached. If the request is persistent, it remains active until it is canceled or this date and time is reached. The default end date is 7 days from the current date.

launchTemplateSpotMarketOptionsRequest_spotInstanceType :: Lens' LaunchTemplateSpotMarketOptionsRequest (Maybe SpotInstanceType)

The Spot Instance request type.

launchTemplateSpotMarketOptionsRequest_maxPrice :: Lens' LaunchTemplateSpotMarketOptionsRequest (Maybe Text)

The maximum hourly price you're willing to pay for the Spot Instances.