Network.AWS.EC2.Types.LocalGatewayRouteTableVpcAssociation

Description

Documentation

data LocalGatewayRouteTableVpcAssociation

Describes an association between a local gateway route table and a VPC.

See: newLocalGatewayRouteTableVpcAssociation smart constructor.

Constructors

LocalGatewayRouteTableVpcAssociation'

Fields

ownerId :: Maybe Text

The AWS account ID that owns the local gateway route table for the association.

localGatewayRouteTableVpcAssociationId :: Maybe Text

The ID of the association.

localGatewayId :: Maybe Text

The ID of the local gateway.

localGatewayRouteTableArn :: Maybe Text

The Amazon Resource Name (ARN) of the local gateway route table for the association.

localGatewayRouteTableId :: Maybe Text

The ID of the local gateway route table.

state :: Maybe Text

The state of the association.

tags :: Maybe [Tag]

The tags assigned to the association.

vpcId :: Maybe Text

The ID of the VPC.

newLocalGatewayRouteTableVpcAssociation :: LocalGatewayRouteTableVpcAssociation

Create a value of LocalGatewayRouteTableVpcAssociation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ownerId:LocalGatewayRouteTableVpcAssociation', localGatewayRouteTableVpcAssociation_ownerId - The AWS account ID that owns the local gateway route table for the association.

$sel:localGatewayRouteTableVpcAssociationId:LocalGatewayRouteTableVpcAssociation', localGatewayRouteTableVpcAssociation_localGatewayRouteTableVpcAssociationId - The ID of the association.

$sel:localGatewayId:LocalGatewayRouteTableVpcAssociation', localGatewayRouteTableVpcAssociation_localGatewayId - The ID of the local gateway.

$sel:localGatewayRouteTableArn:LocalGatewayRouteTableVpcAssociation', localGatewayRouteTableVpcAssociation_localGatewayRouteTableArn - The Amazon Resource Name (ARN) of the local gateway route table for the association.

$sel:localGatewayRouteTableId:LocalGatewayRouteTableVpcAssociation', localGatewayRouteTableVpcAssociation_localGatewayRouteTableId - The ID of the local gateway route table.

$sel:state:LocalGatewayRouteTableVpcAssociation', localGatewayRouteTableVpcAssociation_state - The state of the association.

$sel:tags:LocalGatewayRouteTableVpcAssociation', localGatewayRouteTableVpcAssociation_tags - The tags assigned to the association.

$sel:vpcId:LocalGatewayRouteTableVpcAssociation', localGatewayRouteTableVpcAssociation_vpcId - The ID of the VPC.

localGatewayRouteTableVpcAssociation_ownerId :: Lens' LocalGatewayRouteTableVpcAssociation (Maybe Text)

The AWS account ID that owns the local gateway route table for the association.

localGatewayRouteTableVpcAssociation_localGatewayRouteTableVpcAssociationId :: Lens' LocalGatewayRouteTableVpcAssociation (Maybe Text)

The ID of the association.

localGatewayRouteTableVpcAssociation_localGatewayId :: Lens' LocalGatewayRouteTableVpcAssociation (Maybe Text)

The ID of the local gateway.

localGatewayRouteTableVpcAssociation_localGatewayRouteTableArn :: Lens' LocalGatewayRouteTableVpcAssociation (Maybe Text)

The Amazon Resource Name (ARN) of the local gateway route table for the association.

localGatewayRouteTableVpcAssociation_localGatewayRouteTableId :: Lens' LocalGatewayRouteTableVpcAssociation (Maybe Text)

The ID of the local gateway route table.

localGatewayRouteTableVpcAssociation_state :: Lens' LocalGatewayRouteTableVpcAssociation (Maybe Text)

The state of the association.

localGatewayRouteTableVpcAssociation_tags :: Lens' LocalGatewayRouteTableVpcAssociation (Maybe [Tag])

The tags assigned to the association.

localGatewayRouteTableVpcAssociation_vpcId :: Lens' LocalGatewayRouteTableVpcAssociation (Maybe Text)

The ID of the VPC.