Network.AWS.EC2.Types.ManagedPrefixList

Description

Documentation

data ManagedPrefixList

Describes a managed prefix list.

See: newManagedPrefixList smart constructor.

Constructors

ManagedPrefixList'

Fields

ownerId :: Maybe Text

The ID of the owner of the prefix list.

stateMessage :: Maybe Text

The state message.

maxEntries :: Maybe Int

The maximum number of entries for the prefix list.

prefixListName :: Maybe Text

The name of the prefix list.

version :: Maybe Integer

The version of the prefix list.

prefixListId :: Maybe Text

The ID of the prefix list.

prefixListArn :: Maybe Text

The Amazon Resource Name (ARN) for the prefix list.

state :: Maybe PrefixListState

The current state of the prefix list.

tags :: Maybe [Tag]

The tags for the prefix list.

addressFamily :: Maybe Text

The IP address version.

newManagedPrefixList :: ManagedPrefixList

Create a value of ManagedPrefixList with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ownerId:ManagedPrefixList', managedPrefixList_ownerId - The ID of the owner of the prefix list.

$sel:stateMessage:ManagedPrefixList', managedPrefixList_stateMessage - The state message.

$sel:maxEntries:ManagedPrefixList', managedPrefixList_maxEntries - The maximum number of entries for the prefix list.

$sel:prefixListName:ManagedPrefixList', managedPrefixList_prefixListName - The name of the prefix list.

$sel:version:ManagedPrefixList', managedPrefixList_version - The version of the prefix list.

$sel:prefixListId:ManagedPrefixList', managedPrefixList_prefixListId - The ID of the prefix list.

$sel:prefixListArn:ManagedPrefixList', managedPrefixList_prefixListArn - The Amazon Resource Name (ARN) for the prefix list.

$sel:state:ManagedPrefixList', managedPrefixList_state - The current state of the prefix list.

$sel:tags:ManagedPrefixList', managedPrefixList_tags - The tags for the prefix list.

$sel:addressFamily:ManagedPrefixList', managedPrefixList_addressFamily - The IP address version.

managedPrefixList_ownerId :: Lens' ManagedPrefixList (Maybe Text)

The ID of the owner of the prefix list.

managedPrefixList_stateMessage :: Lens' ManagedPrefixList (Maybe Text)

The state message.

managedPrefixList_maxEntries :: Lens' ManagedPrefixList (Maybe Int)

The maximum number of entries for the prefix list.

managedPrefixList_prefixListName :: Lens' ManagedPrefixList (Maybe Text)

The name of the prefix list.

managedPrefixList_version :: Lens' ManagedPrefixList (Maybe Integer)

The version of the prefix list.

managedPrefixList_prefixListId :: Lens' ManagedPrefixList (Maybe Text)

The ID of the prefix list.

managedPrefixList_prefixListArn :: Lens' ManagedPrefixList (Maybe Text)

The Amazon Resource Name (ARN) for the prefix list.

managedPrefixList_state :: Lens' ManagedPrefixList (Maybe PrefixListState)

The current state of the prefix list.

managedPrefixList_tags :: Lens' ManagedPrefixList (Maybe [Tag])

The tags for the prefix list.

managedPrefixList_addressFamily :: Lens' ManagedPrefixList (Maybe Text)

The IP address version.