Network.AWS.EC2.Types.NatGateway

Description

Documentation

data NatGateway

Describes a NAT gateway.

See: newNatGateway smart constructor.

Constructors

NatGateway'

Fields

natGatewayAddresses :: Maybe [NatGatewayAddress]

Information about the IP addresses and network interface associated with the NAT gateway.

failureMessage :: Maybe Text

If the NAT gateway could not be created, specifies the error message for the failure, that corresponds to the error code.

For InsufficientFreeAddressesInSubnet: "Subnet has insufficient free addresses to create this NAT gateway"

For Gateway.NotAttached: "Network vpc-xxxxxxxx has no Internet gateway attached"

For InvalidAllocationID.NotFound: "Elastic IP address eipalloc-xxxxxxxx could not be associated with this NAT gateway"

For Resource.AlreadyAssociated: "Elastic IP address eipalloc-xxxxxxxx is already associated"

For InternalError: "Network interface eni-xxxxxxxx, created and used internally by this NAT gateway is in an invalid state. Please try again."

For InvalidSubnetID.NotFound: "The specified subnet subnet-xxxxxxxx does not exist or could not be found."

failureCode :: Maybe Text

If the NAT gateway could not be created, specifies the error code for the failure. (InsufficientFreeAddressesInSubnet | Gateway.NotAttached | InvalidAllocationID.NotFound | Resource.AlreadyAssociated | InternalError | InvalidSubnetID.NotFound)

provisionedBandwidth :: Maybe ProvisionedBandwidth

Reserved. If you need to sustain traffic greater than the documented limits, contact us through the Support Center.

state :: Maybe NatGatewayState

The state of the NAT gateway.

pending: The NAT gateway is being created and is not ready to process traffic.

failed: The NAT gateway could not be created. Check the failureCode and failureMessage fields for the reason.

available: The NAT gateway is able to process traffic. This status remains until you delete the NAT gateway, and does not indicate the health of the NAT gateway.

deleting: The NAT gateway is in the process of being terminated and may still be processing traffic.

deleted: The NAT gateway has been terminated and is no longer processing traffic.

deleteTime :: Maybe ISO8601

The date and time the NAT gateway was deleted, if applicable.

createTime :: Maybe ISO8601

The date and time the NAT gateway was created.

tags :: Maybe [Tag]

The tags for the NAT gateway.

subnetId :: Maybe Text

The ID of the subnet in which the NAT gateway is located.

natGatewayId :: Maybe Text

The ID of the NAT gateway.

vpcId :: Maybe Text

The ID of the VPC in which the NAT gateway is located.

connectivityType :: Maybe ConnectivityType

Indicates whether the NAT gateway supports public or private connectivity.

newNatGateway :: NatGateway

Create a value of NatGateway with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:natGatewayAddresses:NatGateway', natGateway_natGatewayAddresses - Information about the IP addresses and network interface associated with the NAT gateway.

$sel:failureMessage:NatGateway', natGateway_failureMessage - If the NAT gateway could not be created, specifies the error message for the failure, that corresponds to the error code.

For InsufficientFreeAddressesInSubnet: "Subnet has insufficient free addresses to create this NAT gateway"

For Gateway.NotAttached: "Network vpc-xxxxxxxx has no Internet gateway attached"

For InvalidAllocationID.NotFound: "Elastic IP address eipalloc-xxxxxxxx could not be associated with this NAT gateway"

For Resource.AlreadyAssociated: "Elastic IP address eipalloc-xxxxxxxx is already associated"

For InternalError: "Network interface eni-xxxxxxxx, created and used internally by this NAT gateway is in an invalid state. Please try again."

For InvalidSubnetID.NotFound: "The specified subnet subnet-xxxxxxxx does not exist or could not be found."

$sel:failureCode:NatGateway', natGateway_failureCode - If the NAT gateway could not be created, specifies the error code for the failure. (InsufficientFreeAddressesInSubnet | Gateway.NotAttached | InvalidAllocationID.NotFound | Resource.AlreadyAssociated | InternalError | InvalidSubnetID.NotFound)

$sel:provisionedBandwidth:NatGateway', natGateway_provisionedBandwidth - Reserved. If you need to sustain traffic greater than the documented limits, contact us through the Support Center.

$sel:state:NatGateway', natGateway_state - The state of the NAT gateway.

pending: The NAT gateway is being created and is not ready to process traffic.

failed: The NAT gateway could not be created. Check the failureCode and failureMessage fields for the reason.

available: The NAT gateway is able to process traffic. This status remains until you delete the NAT gateway, and does not indicate the health of the NAT gateway.

deleting: The NAT gateway is in the process of being terminated and may still be processing traffic.

deleted: The NAT gateway has been terminated and is no longer processing traffic.

$sel:deleteTime:NatGateway', natGateway_deleteTime - The date and time the NAT gateway was deleted, if applicable.

$sel:createTime:NatGateway', natGateway_createTime - The date and time the NAT gateway was created.

$sel:tags:NatGateway', natGateway_tags - The tags for the NAT gateway.

$sel:subnetId:NatGateway', natGateway_subnetId - The ID of the subnet in which the NAT gateway is located.

$sel:natGatewayId:NatGateway', natGateway_natGatewayId - The ID of the NAT gateway.

$sel:vpcId:NatGateway', natGateway_vpcId - The ID of the VPC in which the NAT gateway is located.

$sel:connectivityType:NatGateway', natGateway_connectivityType - Indicates whether the NAT gateway supports public or private connectivity.

natGateway_natGatewayAddresses :: Lens' NatGateway (Maybe [NatGatewayAddress])

Information about the IP addresses and network interface associated with the NAT gateway.

natGateway_failureMessage :: Lens' NatGateway (Maybe Text)

If the NAT gateway could not be created, specifies the error message for the failure, that corresponds to the error code.

For InsufficientFreeAddressesInSubnet: "Subnet has insufficient free addresses to create this NAT gateway"

For Gateway.NotAttached: "Network vpc-xxxxxxxx has no Internet gateway attached"

For InvalidAllocationID.NotFound: "Elastic IP address eipalloc-xxxxxxxx could not be associated with this NAT gateway"

For Resource.AlreadyAssociated: "Elastic IP address eipalloc-xxxxxxxx is already associated"

For InternalError: "Network interface eni-xxxxxxxx, created and used internally by this NAT gateway is in an invalid state. Please try again."

For InvalidSubnetID.NotFound: "The specified subnet subnet-xxxxxxxx does not exist or could not be found."

natGateway_failureCode :: Lens' NatGateway (Maybe Text)

If the NAT gateway could not be created, specifies the error code for the failure. (InsufficientFreeAddressesInSubnet | Gateway.NotAttached | InvalidAllocationID.NotFound | Resource.AlreadyAssociated | InternalError | InvalidSubnetID.NotFound)

natGateway_provisionedBandwidth :: Lens' NatGateway (Maybe ProvisionedBandwidth)

Reserved. If you need to sustain traffic greater than the documented limits, contact us through the Support Center.

natGateway_state :: Lens' NatGateway (Maybe NatGatewayState)

The state of the NAT gateway.

pending: The NAT gateway is being created and is not ready to process traffic.

failed: The NAT gateway could not be created. Check the failureCode and failureMessage fields for the reason.

available: The NAT gateway is able to process traffic. This status remains until you delete the NAT gateway, and does not indicate the health of the NAT gateway.

deleting: The NAT gateway is in the process of being terminated and may still be processing traffic.

deleted: The NAT gateway has been terminated and is no longer processing traffic.

natGateway_deleteTime :: Lens' NatGateway (Maybe UTCTime)

The date and time the NAT gateway was deleted, if applicable.

natGateway_createTime :: Lens' NatGateway (Maybe UTCTime)

The date and time the NAT gateway was created.

natGateway_tags :: Lens' NatGateway (Maybe [Tag])

The tags for the NAT gateway.

natGateway_subnetId :: Lens' NatGateway (Maybe Text)

The ID of the subnet in which the NAT gateway is located.

natGateway_natGatewayId :: Lens' NatGateway (Maybe Text)

The ID of the NAT gateway.

natGateway_vpcId :: Lens' NatGateway (Maybe Text)

The ID of the VPC in which the NAT gateway is located.

natGateway_connectivityType :: Lens' NatGateway (Maybe ConnectivityType)

Indicates whether the NAT gateway supports public or private connectivity.