Network.AWS.EC2.Types.NetworkCardInfo

Description

Documentation

data NetworkCardInfo

Describes the network card support of the instance type.

See: newNetworkCardInfo smart constructor.

Constructors

NetworkCardInfo'

Fields

maximumNetworkInterfaces :: Maybe Int

The maximum number of network interfaces for the network card.

networkCardIndex :: Maybe Int

The index of the network card.

networkPerformance :: Maybe Text

The network performance of the network card.

newNetworkCardInfo :: NetworkCardInfo

Create a value of NetworkCardInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maximumNetworkInterfaces:NetworkCardInfo', networkCardInfo_maximumNetworkInterfaces - The maximum number of network interfaces for the network card.

$sel:networkCardIndex:NetworkCardInfo', networkCardInfo_networkCardIndex - The index of the network card.

$sel:networkPerformance:NetworkCardInfo', networkCardInfo_networkPerformance - The network performance of the network card.

networkCardInfo_maximumNetworkInterfaces :: Lens' NetworkCardInfo (Maybe Int)

The maximum number of network interfaces for the network card.

networkCardInfo_networkCardIndex :: Lens' NetworkCardInfo (Maybe Int)

The index of the network card.

networkCardInfo_networkPerformance :: Lens' NetworkCardInfo (Maybe Text)

The network performance of the network card.