Network.AWS.EC2.Types.NetworkInsightsAnalysis

Description

Documentation

data NetworkInsightsAnalysis

Describes a network insights analysis.

See: newNetworkInsightsAnalysis smart constructor.

Constructors

NetworkInsightsAnalysis'

Fields

statusMessage :: Maybe Text

The status message, if the status is failed.

status :: Maybe AnalysisStatus

The status of the network insights analysis.

networkInsightsAnalysisId :: Maybe Text

The ID of the network insights analysis.

startDate :: Maybe ISO8601

The time the analysis started.

filterInArns :: Maybe [Text]

The Amazon Resource Names (ARN) of the Amazon Web Services resources that the path must traverse.

returnPathComponents :: Maybe [PathComponent]

The components in the path from destination to source.

explanations :: Maybe [Explanation]

The explanations. For more information, see Reachability Analyzer explanation codes.

networkPathFound :: Maybe Bool

Indicates whether the destination is reachable from the source.

networkInsightsAnalysisArn :: Maybe Text

The Amazon Resource Name (ARN) of the network insights analysis.

tags :: Maybe [Tag]

The tags.

networkInsightsPathId :: Maybe Text

The ID of the path.

alternatePathHints :: Maybe [AlternatePathHint]

Potential intermediate components.

forwardPathComponents :: Maybe [PathComponent]

The components in the path from source to destination.

newNetworkInsightsAnalysis :: NetworkInsightsAnalysis

Create a value of NetworkInsightsAnalysis with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:statusMessage:NetworkInsightsAnalysis', networkInsightsAnalysis_statusMessage - The status message, if the status is failed.

$sel:status:NetworkInsightsAnalysis', networkInsightsAnalysis_status - The status of the network insights analysis.

$sel:networkInsightsAnalysisId:NetworkInsightsAnalysis', networkInsightsAnalysis_networkInsightsAnalysisId - The ID of the network insights analysis.

$sel:startDate:NetworkInsightsAnalysis', networkInsightsAnalysis_startDate - The time the analysis started.

$sel:filterInArns:NetworkInsightsAnalysis', networkInsightsAnalysis_filterInArns - The Amazon Resource Names (ARN) of the Amazon Web Services resources that the path must traverse.

$sel:returnPathComponents:NetworkInsightsAnalysis', networkInsightsAnalysis_returnPathComponents - The components in the path from destination to source.

$sel:explanations:NetworkInsightsAnalysis', networkInsightsAnalysis_explanations - The explanations. For more information, see Reachability Analyzer explanation codes.

$sel:networkPathFound:NetworkInsightsAnalysis', networkInsightsAnalysis_networkPathFound - Indicates whether the destination is reachable from the source.

$sel:networkInsightsAnalysisArn:NetworkInsightsAnalysis', networkInsightsAnalysis_networkInsightsAnalysisArn - The Amazon Resource Name (ARN) of the network insights analysis.

$sel:tags:NetworkInsightsAnalysis', networkInsightsAnalysis_tags - The tags.

$sel:networkInsightsPathId:NetworkInsightsAnalysis', networkInsightsAnalysis_networkInsightsPathId - The ID of the path.

$sel:alternatePathHints:NetworkInsightsAnalysis', networkInsightsAnalysis_alternatePathHints - Potential intermediate components.

$sel:forwardPathComponents:NetworkInsightsAnalysis', networkInsightsAnalysis_forwardPathComponents - The components in the path from source to destination.

networkInsightsAnalysis_statusMessage :: Lens' NetworkInsightsAnalysis (Maybe Text)

The status message, if the status is failed.

networkInsightsAnalysis_status :: Lens' NetworkInsightsAnalysis (Maybe AnalysisStatus)

The status of the network insights analysis.

networkInsightsAnalysis_networkInsightsAnalysisId :: Lens' NetworkInsightsAnalysis (Maybe Text)

The ID of the network insights analysis.

networkInsightsAnalysis_startDate :: Lens' NetworkInsightsAnalysis (Maybe UTCTime)

The time the analysis started.

networkInsightsAnalysis_filterInArns :: Lens' NetworkInsightsAnalysis (Maybe [Text])

The Amazon Resource Names (ARN) of the Amazon Web Services resources that the path must traverse.

networkInsightsAnalysis_returnPathComponents :: Lens' NetworkInsightsAnalysis (Maybe [PathComponent])

The components in the path from destination to source.

networkInsightsAnalysis_explanations :: Lens' NetworkInsightsAnalysis (Maybe [Explanation])

The explanations. For more information, see Reachability Analyzer explanation codes.

networkInsightsAnalysis_networkPathFound :: Lens' NetworkInsightsAnalysis (Maybe Bool)

Indicates whether the destination is reachable from the source.

networkInsightsAnalysis_networkInsightsAnalysisArn :: Lens' NetworkInsightsAnalysis (Maybe Text)

The Amazon Resource Name (ARN) of the network insights analysis.

networkInsightsAnalysis_tags :: Lens' NetworkInsightsAnalysis (Maybe [Tag])

The tags.

networkInsightsAnalysis_networkInsightsPathId :: Lens' NetworkInsightsAnalysis (Maybe Text)

The ID of the path.

networkInsightsAnalysis_alternatePathHints :: Lens' NetworkInsightsAnalysis (Maybe [AlternatePathHint])

Potential intermediate components.

networkInsightsAnalysis_forwardPathComponents :: Lens' NetworkInsightsAnalysis (Maybe [PathComponent])

The components in the path from source to destination.