Network.AWS.EC2.Types.NetworkInsightsPath

Description

Documentation

data NetworkInsightsPath

Describes a path.

See: newNetworkInsightsPath smart constructor.

Constructors

NetworkInsightsPath'

Fields

createdDate :: Maybe ISO8601

The time stamp when the path was created.

source :: Maybe Text

The Amazon Web Services resource that is the source of the path.

destinationIp :: Maybe Text

The IP address of the Amazon Web Services resource that is the destination of the path.

sourceIp :: Maybe Text

The IP address of the Amazon Web Services resource that is the source of the path.

destination :: Maybe Text

The Amazon Web Services resource that is the destination of the path.

tags :: Maybe [Tag]

The tags associated with the path.

protocol :: Maybe Protocol

The protocol.

networkInsightsPathId :: Maybe Text

The ID of the path.

networkInsightsPathArn :: Maybe Text

The Amazon Resource Name (ARN) of the path.

destinationPort :: Maybe Int

The destination port.

newNetworkInsightsPath :: NetworkInsightsPath

Create a value of NetworkInsightsPath with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createdDate:NetworkInsightsPath', networkInsightsPath_createdDate - The time stamp when the path was created.

$sel:source:NetworkInsightsPath', networkInsightsPath_source - The Amazon Web Services resource that is the source of the path.

$sel:destinationIp:NetworkInsightsPath', networkInsightsPath_destinationIp - The IP address of the Amazon Web Services resource that is the destination of the path.

$sel:sourceIp:NetworkInsightsPath', networkInsightsPath_sourceIp - The IP address of the Amazon Web Services resource that is the source of the path.

$sel:destination:NetworkInsightsPath', networkInsightsPath_destination - The Amazon Web Services resource that is the destination of the path.

$sel:tags:NetworkInsightsPath', networkInsightsPath_tags - The tags associated with the path.

$sel:protocol:NetworkInsightsPath', networkInsightsPath_protocol - The protocol.

$sel:networkInsightsPathId:NetworkInsightsPath', networkInsightsPath_networkInsightsPathId - The ID of the path.

$sel:networkInsightsPathArn:NetworkInsightsPath', networkInsightsPath_networkInsightsPathArn - The Amazon Resource Name (ARN) of the path.

$sel:destinationPort:NetworkInsightsPath', networkInsightsPath_destinationPort - The destination port.

networkInsightsPath_createdDate :: Lens' NetworkInsightsPath (Maybe UTCTime)

The time stamp when the path was created.

networkInsightsPath_source :: Lens' NetworkInsightsPath (Maybe Text)

The Amazon Web Services resource that is the source of the path.

networkInsightsPath_destinationIp :: Lens' NetworkInsightsPath (Maybe Text)

The IP address of the Amazon Web Services resource that is the destination of the path.

networkInsightsPath_sourceIp :: Lens' NetworkInsightsPath (Maybe Text)

The IP address of the Amazon Web Services resource that is the source of the path.

networkInsightsPath_destination :: Lens' NetworkInsightsPath (Maybe Text)

The Amazon Web Services resource that is the destination of the path.

networkInsightsPath_tags :: Lens' NetworkInsightsPath (Maybe [Tag])

The tags associated with the path.

networkInsightsPath_protocol :: Lens' NetworkInsightsPath (Maybe Protocol)

The protocol.

networkInsightsPath_networkInsightsPathId :: Lens' NetworkInsightsPath (Maybe Text)

The ID of the path.

networkInsightsPath_networkInsightsPathArn :: Lens' NetworkInsightsPath (Maybe Text)

The Amazon Resource Name (ARN) of the path.

networkInsightsPath_destinationPort :: Lens' NetworkInsightsPath (Maybe Int)

The destination port.