Network.AWS.EC2.Types.NetworkInterfaceAttachment

Description

Documentation

data NetworkInterfaceAttachment

Describes a network interface attachment.

See: newNetworkInterfaceAttachment smart constructor.

Constructors

NetworkInterfaceAttachment'

Fields

status :: Maybe AttachmentStatus

The attachment state.

instanceId :: Maybe Text

The ID of the instance.

attachTime :: Maybe ISO8601

The timestamp indicating when the attachment initiated.

deleteOnTermination :: Maybe Bool

Indicates whether the network interface is deleted when the instance is terminated.

attachmentId :: Maybe Text

The ID of the network interface attachment.

networkCardIndex :: Maybe Int

The index of the network card.

deviceIndex :: Maybe Int

The device index of the network interface attachment on the instance.

instanceOwnerId :: Maybe Text

The Amazon Web Services account ID of the owner of the instance.

newNetworkInterfaceAttachment :: NetworkInterfaceAttachment

Create a value of NetworkInterfaceAttachment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:NetworkInterfaceAttachment', networkInterfaceAttachment_status - The attachment state.

$sel:instanceId:NetworkInterfaceAttachment', networkInterfaceAttachment_instanceId - The ID of the instance.

$sel:attachTime:NetworkInterfaceAttachment', networkInterfaceAttachment_attachTime - The timestamp indicating when the attachment initiated.

$sel:deleteOnTermination:NetworkInterfaceAttachment', networkInterfaceAttachment_deleteOnTermination - Indicates whether the network interface is deleted when the instance is terminated.

$sel:attachmentId:NetworkInterfaceAttachment', networkInterfaceAttachment_attachmentId - The ID of the network interface attachment.

$sel:networkCardIndex:NetworkInterfaceAttachment', networkInterfaceAttachment_networkCardIndex - The index of the network card.

$sel:deviceIndex:NetworkInterfaceAttachment', networkInterfaceAttachment_deviceIndex - The device index of the network interface attachment on the instance.

$sel:instanceOwnerId:NetworkInterfaceAttachment', networkInterfaceAttachment_instanceOwnerId - The Amazon Web Services account ID of the owner of the instance.

networkInterfaceAttachment_status :: Lens' NetworkInterfaceAttachment (Maybe AttachmentStatus)

The attachment state.

networkInterfaceAttachment_instanceId :: Lens' NetworkInterfaceAttachment (Maybe Text)

The ID of the instance.

networkInterfaceAttachment_attachTime :: Lens' NetworkInterfaceAttachment (Maybe UTCTime)

The timestamp indicating when the attachment initiated.

networkInterfaceAttachment_deleteOnTermination :: Lens' NetworkInterfaceAttachment (Maybe Bool)

Indicates whether the network interface is deleted when the instance is terminated.

networkInterfaceAttachment_attachmentId :: Lens' NetworkInterfaceAttachment (Maybe Text)

The ID of the network interface attachment.

networkInterfaceAttachment_networkCardIndex :: Lens' NetworkInterfaceAttachment (Maybe Int)

The index of the network card.

networkInterfaceAttachment_deviceIndex :: Lens' NetworkInterfaceAttachment (Maybe Int)

The device index of the network interface attachment on the instance.

networkInterfaceAttachment_instanceOwnerId :: Lens' NetworkInterfaceAttachment (Maybe Text)

The Amazon Web Services account ID of the owner of the instance.