Network.AWS.EC2.Types.Placement

Description

Documentation

data Placement

Describes the placement of an instance.

See: newPlacement smart constructor.

Constructors

Placement'

Fields

affinity :: Maybe Text

The affinity setting for the instance on the Dedicated Host. This parameter is not supported for the ImportInstance command.

This parameter is not supported by CreateFleet.

hostId :: Maybe Text

The ID of the Dedicated Host on which the instance resides. This parameter is not supported for the ImportInstance command.

This parameter is not supported by CreateFleet.

partitionNumber :: Maybe Int

The number of the partition the instance is in. Valid only if the placement group strategy is set to partition.

This parameter is not supported by CreateFleet.

spreadDomain :: Maybe Text

Reserved for future use.

This parameter is not supported by CreateFleet.

availabilityZone :: Maybe Text

The Availability Zone of the instance.

If not specified, an Availability Zone will be automatically chosen for you based on the load balancing criteria for the Region.

This parameter is not supported by CreateFleet.

tenancy :: Maybe Tenancy

The tenancy of the instance (if the instance is running in a VPC). An instance with a tenancy of dedicated runs on single-tenant hardware. The host tenancy is not supported for the ImportInstance command.

This parameter is not supported by CreateFleet.

T3 instances that use the unlimited CPU credit option do not support host tenancy.

groupName :: Maybe Text

The name of the placement group the instance is in.

hostResourceGroupArn :: Maybe Text

The ARN of the host resource group in which to launch the instances. If you specify a host resource group ARN, omit the Tenancy parameter or set it to host.

This parameter is not supported by CreateFleet.

newPlacement :: Placement

Create a value of Placement with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:affinity:Placement', placement_affinity - The affinity setting for the instance on the Dedicated Host. This parameter is not supported for the ImportInstance command.

This parameter is not supported by CreateFleet.

$sel:hostId:Placement', placement_hostId - The ID of the Dedicated Host on which the instance resides. This parameter is not supported for the ImportInstance command.

This parameter is not supported by CreateFleet.

$sel:partitionNumber:Placement', placement_partitionNumber - The number of the partition the instance is in. Valid only if the placement group strategy is set to partition.

This parameter is not supported by CreateFleet.

$sel:spreadDomain:Placement', placement_spreadDomain - Reserved for future use.

This parameter is not supported by CreateFleet.

$sel:availabilityZone:Placement', placement_availabilityZone - The Availability Zone of the instance.

If not specified, an Availability Zone will be automatically chosen for you based on the load balancing criteria for the Region.

This parameter is not supported by CreateFleet.

$sel:tenancy:Placement', placement_tenancy - The tenancy of the instance (if the instance is running in a VPC). An instance with a tenancy of dedicated runs on single-tenant hardware. The host tenancy is not supported for the ImportInstance command.

This parameter is not supported by CreateFleet.

T3 instances that use the unlimited CPU credit option do not support host tenancy.

$sel:groupName:Placement', placement_groupName - The name of the placement group the instance is in.

$sel:hostResourceGroupArn:Placement', placement_hostResourceGroupArn - The ARN of the host resource group in which to launch the instances. If you specify a host resource group ARN, omit the Tenancy parameter or set it to host.

This parameter is not supported by CreateFleet.

placement_affinity :: Lens' Placement (Maybe Text)

The affinity setting for the instance on the Dedicated Host. This parameter is not supported for the ImportInstance command.

This parameter is not supported by CreateFleet.

placement_hostId :: Lens' Placement (Maybe Text)

The ID of the Dedicated Host on which the instance resides. This parameter is not supported for the ImportInstance command.

This parameter is not supported by CreateFleet.

placement_partitionNumber :: Lens' Placement (Maybe Int)

The number of the partition the instance is in. Valid only if the placement group strategy is set to partition.

This parameter is not supported by CreateFleet.

placement_spreadDomain :: Lens' Placement (Maybe Text)

Reserved for future use.

This parameter is not supported by CreateFleet.

placement_availabilityZone :: Lens' Placement (Maybe Text)

The Availability Zone of the instance.

If not specified, an Availability Zone will be automatically chosen for you based on the load balancing criteria for the Region.

This parameter is not supported by CreateFleet.

placement_tenancy :: Lens' Placement (Maybe Tenancy)

The tenancy of the instance (if the instance is running in a VPC). An instance with a tenancy of dedicated runs on single-tenant hardware. The host tenancy is not supported for the ImportInstance command.

This parameter is not supported by CreateFleet.

T3 instances that use the unlimited CPU credit option do not support host tenancy.

placement_groupName :: Lens' Placement (Maybe Text)

The name of the placement group the instance is in.

placement_hostResourceGroupArn :: Lens' Placement (Maybe Text)

The ARN of the host resource group in which to launch the instances. If you specify a host resource group ARN, omit the Tenancy parameter or set it to host.

This parameter is not supported by CreateFleet.