Network.AWS.EC2.Types.PriceSchedule

Description

Documentation

data PriceSchedule

Describes the price for a Reserved Instance.

See: newPriceSchedule smart constructor.

Constructors

PriceSchedule'

Fields

currencyCode :: Maybe CurrencyCodeValues

The currency for transacting the Reserved Instance resale. At this time, the only supported currency is USD.

term :: Maybe Integer

The number of months remaining in the reservation. For example, 2 is the second to the last month before the capacity reservation expires.

active :: Maybe Bool

The current price schedule, as determined by the term remaining for the Reserved Instance in the listing.

A specific price schedule is always in effect, but only one price schedule can be active at any time. Take, for example, a Reserved Instance listing that has five months remaining in its term. When you specify price schedules for five months and two months, this means that schedule 1, covering the first three months of the remaining term, will be active during months 5, 4, and 3. Then schedule 2, covering the last two months of the term, will be active for months 2 and 1.

price :: Maybe Double

The fixed price for the term.

newPriceSchedule :: PriceSchedule

Create a value of PriceSchedule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:currencyCode:PriceSchedule', priceSchedule_currencyCode - The currency for transacting the Reserved Instance resale. At this time, the only supported currency is USD.

$sel:term:PriceSchedule', priceSchedule_term - The number of months remaining in the reservation. For example, 2 is the second to the last month before the capacity reservation expires.

$sel:active:PriceSchedule', priceSchedule_active - The current price schedule, as determined by the term remaining for the Reserved Instance in the listing.

A specific price schedule is always in effect, but only one price schedule can be active at any time. Take, for example, a Reserved Instance listing that has five months remaining in its term. When you specify price schedules for five months and two months, this means that schedule 1, covering the first three months of the remaining term, will be active during months 5, 4, and 3. Then schedule 2, covering the last two months of the term, will be active for months 2 and 1.

$sel:price:PriceSchedule', priceSchedule_price - The fixed price for the term.

priceSchedule_currencyCode :: Lens' PriceSchedule (Maybe CurrencyCodeValues)

The currency for transacting the Reserved Instance resale. At this time, the only supported currency is USD.

priceSchedule_term :: Lens' PriceSchedule (Maybe Integer)

The number of months remaining in the reservation. For example, 2 is the second to the last month before the capacity reservation expires.

priceSchedule_active :: Lens' PriceSchedule (Maybe Bool)

The current price schedule, as determined by the term remaining for the Reserved Instance in the listing.

A specific price schedule is always in effect, but only one price schedule can be active at any time. Take, for example, a Reserved Instance listing that has five months remaining in its term. When you specify price schedules for five months and two months, this means that schedule 1, covering the first three months of the remaining term, will be active during months 5, 4, and 3. Then schedule 2, covering the last two months of the term, will be active for months 2 and 1.

priceSchedule_price :: Lens' PriceSchedule (Maybe Double)

The fixed price for the term.