Network.AWS.EC2.Types.PriceScheduleSpecification

Description

Documentation

data PriceScheduleSpecification

Describes the price for a Reserved Instance.

See: newPriceScheduleSpecification smart constructor.

Constructors

PriceScheduleSpecification'

Fields

currencyCode :: Maybe CurrencyCodeValues

The currency for transacting the Reserved Instance resale. At this time, the only supported currency is USD.

term :: Maybe Integer

The number of months remaining in the reservation. For example, 2 is the second to the last month before the capacity reservation expires.

price :: Maybe Double

The fixed price for the term.

newPriceScheduleSpecification :: PriceScheduleSpecification

Create a value of PriceScheduleSpecification with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:currencyCode:PriceScheduleSpecification', priceScheduleSpecification_currencyCode - The currency for transacting the Reserved Instance resale. At this time, the only supported currency is USD.

$sel:term:PriceScheduleSpecification', priceScheduleSpecification_term - The number of months remaining in the reservation. For example, 2 is the second to the last month before the capacity reservation expires.

$sel:price:PriceScheduleSpecification', priceScheduleSpecification_price - The fixed price for the term.

priceScheduleSpecification_currencyCode :: Lens' PriceScheduleSpecification (Maybe CurrencyCodeValues)

The currency for transacting the Reserved Instance resale. At this time, the only supported currency is USD.

priceScheduleSpecification_term :: Lens' PriceScheduleSpecification (Maybe Integer)

The number of months remaining in the reservation. For example, 2 is the second to the last month before the capacity reservation expires.

priceScheduleSpecification_price :: Lens' PriceScheduleSpecification (Maybe Double)

The fixed price for the term.