Network.AWS.EC2.Types.PrivateDnsNameConfiguration

Description

Documentation

data PrivateDnsNameConfiguration

Information about the private DNS name for the service endpoint. For more information about these parameters, see VPC Endpoint Service Private DNS Name Verification in the Amazon Virtual Private Cloud User Guide.

See: newPrivateDnsNameConfiguration smart constructor.

Constructors

PrivateDnsNameConfiguration'

Fields

name :: Maybe Text

The name of the record subdomain the service provider needs to create. The service provider adds the value text to the name.

state :: Maybe DnsNameState

The verification state of the VPC endpoint service.

Consumers of the endpoint service can use the private name only when

the state is verified.

value :: Maybe Text

The value the service provider adds to the private DNS name domain record before verification.

type' :: Maybe Text

The endpoint service verification type, for example TXT.

newPrivateDnsNameConfiguration :: PrivateDnsNameConfiguration

Create a value of PrivateDnsNameConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:PrivateDnsNameConfiguration', privateDnsNameConfiguration_name - The name of the record subdomain the service provider needs to create. The service provider adds the value text to the name.

$sel:state:PrivateDnsNameConfiguration', privateDnsNameConfiguration_state - The verification state of the VPC endpoint service.

Consumers of the endpoint service can use the private name only when

the state is verified.

$sel:value:PrivateDnsNameConfiguration', privateDnsNameConfiguration_value - The value the service provider adds to the private DNS name domain record before verification.

$sel:type':PrivateDnsNameConfiguration', privateDnsNameConfiguration_type - The endpoint service verification type, for example TXT.

privateDnsNameConfiguration_name :: Lens' PrivateDnsNameConfiguration (Maybe Text)

The name of the record subdomain the service provider needs to create. The service provider adds the value text to the name.

privateDnsNameConfiguration_state :: Lens' PrivateDnsNameConfiguration (Maybe DnsNameState)

The verification state of the VPC endpoint service.

Consumers of the endpoint service can use the private name only when

the state is verified.

privateDnsNameConfiguration_value :: Lens' PrivateDnsNameConfiguration (Maybe Text)

The value the service provider adds to the private DNS name domain record before verification.

privateDnsNameConfiguration_type :: Lens' PrivateDnsNameConfiguration (Maybe Text)

The endpoint service verification type, for example TXT.