Network.AWS.EC2.Types.ProvisionedBandwidth

Description

Documentation

data ProvisionedBandwidth

Reserved. If you need to sustain traffic greater than the documented limits, contact us through the Support Center.

See: newProvisionedBandwidth smart constructor.

Constructors

ProvisionedBandwidth'

Fields

status :: Maybe Text

Reserved. If you need to sustain traffic greater than the documented limits, contact us through the Support Center.

provisionTime :: Maybe ISO8601

Reserved. If you need to sustain traffic greater than the documented limits, contact us through the Support Center.

requestTime :: Maybe ISO8601

Reserved. If you need to sustain traffic greater than the documented limits, contact us through the Support Center.

requested :: Maybe Text

Reserved. If you need to sustain traffic greater than the documented limits, contact us through the Support Center.

provisioned :: Maybe Text

Reserved. If you need to sustain traffic greater than the documented limits, contact us through the Support Center.

newProvisionedBandwidth :: ProvisionedBandwidth

Create a value of ProvisionedBandwidth with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ProvisionedBandwidth', provisionedBandwidth_status - Reserved. If you need to sustain traffic greater than the documented limits, contact us through the Support Center.

$sel:provisionTime:ProvisionedBandwidth', provisionedBandwidth_provisionTime - Reserved. If you need to sustain traffic greater than the documented limits, contact us through the Support Center.

$sel:requestTime:ProvisionedBandwidth', provisionedBandwidth_requestTime - Reserved. If you need to sustain traffic greater than the documented limits, contact us through the Support Center.

$sel:requested:ProvisionedBandwidth', provisionedBandwidth_requested - Reserved. If you need to sustain traffic greater than the documented limits, contact us through the Support Center.

$sel:provisioned:ProvisionedBandwidth', provisionedBandwidth_provisioned - Reserved. If you need to sustain traffic greater than the documented limits, contact us through the Support Center.

provisionedBandwidth_status :: Lens' ProvisionedBandwidth (Maybe Text)

Reserved. If you need to sustain traffic greater than the documented limits, contact us through the Support Center.

provisionedBandwidth_provisionTime :: Lens' ProvisionedBandwidth (Maybe UTCTime)

Reserved. If you need to sustain traffic greater than the documented limits, contact us through the Support Center.

provisionedBandwidth_requestTime :: Lens' ProvisionedBandwidth (Maybe UTCTime)

Reserved. If you need to sustain traffic greater than the documented limits, contact us through the Support Center.

provisionedBandwidth_requested :: Lens' ProvisionedBandwidth (Maybe Text)

Reserved. If you need to sustain traffic greater than the documented limits, contact us through the Support Center.

provisionedBandwidth_provisioned :: Lens' ProvisionedBandwidth (Maybe Text)

Reserved. If you need to sustain traffic greater than the documented limits, contact us through the Support Center.