Network.AWS.EC2.Types.Purchase

Description

Documentation

data Purchase

Describes the result of the purchase.

See: newPurchase smart constructor.

Constructors

Purchase'

Fields

hostIdSet :: Maybe [Text]

The IDs of the Dedicated Hosts associated with the reservation.

instanceFamily :: Maybe Text

The instance family on the Dedicated Host that the reservation can be associated with.

paymentOption :: Maybe PaymentOption

The payment option for the reservation.

upfrontPrice :: Maybe Text

The upfront price of the reservation.

duration :: Maybe Int

The duration of the reservation's term in seconds.

hostReservationId :: Maybe Text

The ID of the reservation.

currencyCode :: Maybe CurrencyCodeValues

The currency in which the UpfrontPrice and HourlyPrice amounts are specified. At this time, the only supported currency is USD.

hourlyPrice :: Maybe Text

The hourly price of the reservation per hour.

newPurchase :: Purchase

Create a value of Purchase with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hostIdSet:Purchase', purchase_hostIdSet - The IDs of the Dedicated Hosts associated with the reservation.

$sel:instanceFamily:Purchase', purchase_instanceFamily - The instance family on the Dedicated Host that the reservation can be associated with.

$sel:paymentOption:Purchase', purchase_paymentOption - The payment option for the reservation.

$sel:upfrontPrice:Purchase', purchase_upfrontPrice - The upfront price of the reservation.

$sel:duration:Purchase', purchase_duration - The duration of the reservation's term in seconds.

$sel:hostReservationId:Purchase', purchase_hostReservationId - The ID of the reservation.

$sel:currencyCode:Purchase', purchase_currencyCode - The currency in which the UpfrontPrice and HourlyPrice amounts are specified. At this time, the only supported currency is USD.

$sel:hourlyPrice:Purchase', purchase_hourlyPrice - The hourly price of the reservation per hour.

purchase_hostIdSet :: Lens' Purchase (Maybe [Text])

The IDs of the Dedicated Hosts associated with the reservation.

purchase_instanceFamily :: Lens' Purchase (Maybe Text)

The instance family on the Dedicated Host that the reservation can be associated with.

purchase_paymentOption :: Lens' Purchase (Maybe PaymentOption)

The payment option for the reservation.

purchase_upfrontPrice :: Lens' Purchase (Maybe Text)

The upfront price of the reservation.

purchase_duration :: Lens' Purchase (Maybe Int)

The duration of the reservation's term in seconds.

purchase_hostReservationId :: Lens' Purchase (Maybe Text)

The ID of the reservation.

purchase_currencyCode :: Lens' Purchase (Maybe CurrencyCodeValues)

The currency in which the UpfrontPrice and HourlyPrice amounts are specified. At this time, the only supported currency is USD.

purchase_hourlyPrice :: Lens' Purchase (Maybe Text)

The hourly price of the reservation per hour.