Network.AWS.EC2.Types.ReferencedSecurityGroup

Description

Documentation

data ReferencedSecurityGroup

Describes the security group that is referenced in the security group rule.

See: newReferencedSecurityGroup smart constructor.

Constructors

ReferencedSecurityGroup'

Fields

vpcPeeringConnectionId :: Maybe Text

The ID of the VPC peering connection.

groupId :: Maybe Text

The ID of the security group.

userId :: Maybe Text

The Amazon Web Services account ID.

peeringStatus :: Maybe Text

The status of a VPC peering connection, if applicable.

vpcId :: Maybe Text

The ID of the VPC.

newReferencedSecurityGroup :: ReferencedSecurityGroup

Create a value of ReferencedSecurityGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpcPeeringConnectionId:ReferencedSecurityGroup', referencedSecurityGroup_vpcPeeringConnectionId - The ID of the VPC peering connection.

$sel:groupId:ReferencedSecurityGroup', referencedSecurityGroup_groupId - The ID of the security group.

$sel:userId:ReferencedSecurityGroup', referencedSecurityGroup_userId - The Amazon Web Services account ID.

$sel:peeringStatus:ReferencedSecurityGroup', referencedSecurityGroup_peeringStatus - The status of a VPC peering connection, if applicable.

$sel:vpcId:ReferencedSecurityGroup', referencedSecurityGroup_vpcId - The ID of the VPC.

referencedSecurityGroup_vpcPeeringConnectionId :: Lens' ReferencedSecurityGroup (Maybe Text)

The ID of the VPC peering connection.

referencedSecurityGroup_groupId :: Lens' ReferencedSecurityGroup (Maybe Text)

The ID of the security group.

referencedSecurityGroup_userId :: Lens' ReferencedSecurityGroup (Maybe Text)

The Amazon Web Services account ID.

referencedSecurityGroup_peeringStatus :: Lens' ReferencedSecurityGroup (Maybe Text)

The status of a VPC peering connection, if applicable.

referencedSecurityGroup_vpcId :: Lens' ReferencedSecurityGroup (Maybe Text)

The ID of the VPC.