Network.AWS.EC2.Types.ReplaceRootVolumeTask

Description

Documentation

data ReplaceRootVolumeTask

Information about a root volume replacement task.

See: newReplaceRootVolumeTask smart constructor.

Constructors

ReplaceRootVolumeTask'

Fields

instanceId :: Maybe Text

The ID of the instance for which the root volume replacement task was created.

startTime :: Maybe Text

The time the task was started.

replaceRootVolumeTaskId :: Maybe Text

The ID of the root volume replacement task.

completeTime :: Maybe Text

The time the task completed.

tags :: Maybe [Tag]

The tags assigned to the task.

taskState :: Maybe ReplaceRootVolumeTaskState

The state of the task. The task can be in one of the following states:

pending - the replacement volume is being created.

in-progress - the original volume is being detached and the replacement volume is being attached.

succeeded - the replacement volume has been successfully attached to the instance and the instance is available.

failing - the replacement task is in the process of failing.

failed - the replacement task has failed but the original root volume is still attached.

failing-detached - the replacement task is in the process of failing. The instance might have no root volume attached.

failed-detached - the replacement task has failed and the instance has no root volume attached.

newReplaceRootVolumeTask :: ReplaceRootVolumeTask

Create a value of ReplaceRootVolumeTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceId:ReplaceRootVolumeTask', replaceRootVolumeTask_instanceId - The ID of the instance for which the root volume replacement task was created.

$sel:startTime:ReplaceRootVolumeTask', replaceRootVolumeTask_startTime - The time the task was started.

$sel:replaceRootVolumeTaskId:ReplaceRootVolumeTask', replaceRootVolumeTask_replaceRootVolumeTaskId - The ID of the root volume replacement task.

$sel:completeTime:ReplaceRootVolumeTask', replaceRootVolumeTask_completeTime - The time the task completed.

$sel:tags:ReplaceRootVolumeTask', replaceRootVolumeTask_tags - The tags assigned to the task.

$sel:taskState:ReplaceRootVolumeTask', replaceRootVolumeTask_taskState - The state of the task. The task can be in one of the following states:

pending - the replacement volume is being created.

in-progress - the original volume is being detached and the replacement volume is being attached.

succeeded - the replacement volume has been successfully attached to the instance and the instance is available.

failing - the replacement task is in the process of failing.

failed - the replacement task has failed but the original root volume is still attached.

failing-detached - the replacement task is in the process of failing. The instance might have no root volume attached.

failed-detached - the replacement task has failed and the instance has no root volume attached.

replaceRootVolumeTask_instanceId :: Lens' ReplaceRootVolumeTask (Maybe Text)

The ID of the instance for which the root volume replacement task was created.

replaceRootVolumeTask_startTime :: Lens' ReplaceRootVolumeTask (Maybe Text)

The time the task was started.

replaceRootVolumeTask_replaceRootVolumeTaskId :: Lens' ReplaceRootVolumeTask (Maybe Text)

The ID of the root volume replacement task.

replaceRootVolumeTask_completeTime :: Lens' ReplaceRootVolumeTask (Maybe Text)

The time the task completed.

replaceRootVolumeTask_tags :: Lens' ReplaceRootVolumeTask (Maybe [Tag])

The tags assigned to the task.

replaceRootVolumeTask_taskState :: Lens' ReplaceRootVolumeTask (Maybe ReplaceRootVolumeTaskState)

The state of the task. The task can be in one of the following states:

pending - the replacement volume is being created.

in-progress - the original volume is being detached and the replacement volume is being attached.

succeeded - the replacement volume has been successfully attached to the instance and the instance is available.

failing - the replacement task is in the process of failing.

failed - the replacement task has failed but the original root volume is still attached.

failing-detached - the replacement task is in the process of failing. The instance might have no root volume attached.

failed-detached - the replacement task has failed and the instance has no root volume attached.