Network.AWS.EC2.Types.RequestLaunchTemplateData

Description

Documentation

data RequestLaunchTemplateData

The information to include in the launch template.

See: newRequestLaunchTemplateData smart constructor.

Constructors

RequestLaunchTemplateData'

Fields

securityGroupIds :: Maybe [Text]

One or more security group IDs. You can create a security group using CreateSecurityGroup. You cannot specify both a security group ID and security name in the same request.

tagSpecifications :: Maybe [LaunchTemplateTagSpecificationRequest]

The tags to apply to the resources during launch. You can only tag instances and volumes on launch. The specified tags are applied to all instances or volumes that are created during launch. To tag a resource after it has been created, see CreateTags.

elasticGpuSpecifications :: Maybe [ElasticGpuSpecification]

An elastic GPU to associate with the instance.

capacityReservationSpecification :: Maybe LaunchTemplateCapacityReservationSpecificationRequest

The Capacity Reservation targeting option. If you do not specify this parameter, the instance's Capacity Reservation preference defaults to open, which enables it to run in any open Capacity Reservation that has matching attributes (instance type, platform, Availability Zone).

ebsOptimized :: Maybe Bool

Indicates whether the instance is optimized for Amazon EBS I/O. This optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal Amazon EBS I/O performance. This optimization isn't available with all instance types. Additional usage charges apply when using an EBS-optimized instance.

placement :: Maybe LaunchTemplatePlacementRequest

The placement for the instance.

userData :: Maybe Text

The user data to make available to the instance. You must provide base64-encoded text. User data is limited to 16 KB. For more information, see Running Commands on Your Linux Instance at Launch (Linux) or Adding User Data (Windows).

If you are creating the launch template for use with Batch, the user data must be provided in the MIME multi-part archive format. For more information, see Amazon EC2 user data in launch templates in the Batch User Guide.

instanceType :: Maybe InstanceType

The instance type. For more information, see Instance Types in the Amazon Elastic Compute Cloud User Guide.

ramDiskId :: Maybe Text

The ID of the RAM disk.

We recommend that you use PV-GRUB instead of kernels and RAM disks. For more information, see User Provided Kernels in the Amazon Elastic Compute Cloud User Guide.

creditSpecification :: Maybe CreditSpecificationRequest

The credit option for CPU usage of the instance. Valid for T2, T3, or T3a instances only.

licenseSpecifications :: Maybe [LaunchTemplateLicenseConfigurationRequest]

The license configurations.

instanceMarketOptions :: Maybe LaunchTemplateInstanceMarketOptionsRequest

The market (purchasing) option for the instances.

instanceInitiatedShutdownBehavior :: Maybe ShutdownBehavior

Indicates whether an instance stops or terminates when you initiate shutdown from the instance (using the operating system command for system shutdown).

Default: stop

elasticInferenceAccelerators :: Maybe [LaunchTemplateElasticInferenceAccelerator]

The elastic inference accelerator for the instance.

imageId :: Maybe Text

The ID of the AMI.

securityGroups :: Maybe [Text]

EC2-Classic, default VPC

One or more security group names. For a nondefault VPC, you must use security group IDs instead. You cannot specify both a security group ID and security name in the same request.

iamInstanceProfile :: Maybe LaunchTemplateIamInstanceProfileSpecificationRequest

The name or Amazon Resource Name (ARN) of an IAM instance profile.

hibernationOptions :: Maybe LaunchTemplateHibernationOptionsRequest

Indicates whether an instance is enabled for hibernation. This parameter is valid only if the instance meets the hibernation prerequisites. For more information, see Hibernate your instance in the Amazon Elastic Compute Cloud User Guide.

monitoring :: Maybe LaunchTemplatesMonitoringRequest

The monitoring for the instance.

blockDeviceMappings :: Maybe [LaunchTemplateBlockDeviceMappingRequest]

The block device mapping.

enclaveOptions :: Maybe LaunchTemplateEnclaveOptionsRequest

Indicates whether the instance is enabled for Amazon Web Services Nitro Enclaves. For more information, see What is Amazon Web Services Nitro Enclaves? in the Amazon Web Services Nitro Enclaves User Guide.

You can't enable Amazon Web Services Nitro Enclaves and hibernation on the same instance.

kernelId :: Maybe Text

The ID of the kernel.

We recommend that you use PV-GRUB instead of kernels and RAM disks. For more information, see User Provided Kernels in the Amazon Elastic Compute Cloud User Guide.

cpuOptions :: Maybe LaunchTemplateCpuOptionsRequest

The CPU options for the instance. For more information, see Optimizing CPU Options in the Amazon Elastic Compute Cloud User Guide.

disableApiTermination :: Maybe Bool

If you set this parameter to true, you can't terminate the instance using the Amazon EC2 console, CLI, or API; otherwise, you can. To change this attribute after launch, use ModifyInstanceAttribute. Alternatively, if you set InstanceInitiatedShutdownBehavior to terminate, you can terminate the instance by running the shutdown command from the instance.

networkInterfaces :: Maybe [LaunchTemplateInstanceNetworkInterfaceSpecificationRequest]

One or more network interfaces. If you specify a network interface, you must specify any security groups and subnets as part of the network interface.

keyName :: Maybe Text

The name of the key pair. You can create a key pair using CreateKeyPair or ImportKeyPair.

If you do not specify a key pair, you can't connect to the instance unless you choose an AMI that is configured to allow users another way to log in.

metadataOptions :: Maybe LaunchTemplateInstanceMetadataOptionsRequest

The metadata options for the instance. For more information, see Instance metadata and user data in the Amazon Elastic Compute Cloud User Guide.

newRequestLaunchTemplateData :: RequestLaunchTemplateData

Create a value of RequestLaunchTemplateData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:securityGroupIds:RequestLaunchTemplateData', requestLaunchTemplateData_securityGroupIds - One or more security group IDs. You can create a security group using CreateSecurityGroup. You cannot specify both a security group ID and security name in the same request.

$sel:tagSpecifications:RequestLaunchTemplateData', requestLaunchTemplateData_tagSpecifications - The tags to apply to the resources during launch. You can only tag instances and volumes on launch. The specified tags are applied to all instances or volumes that are created during launch. To tag a resource after it has been created, see CreateTags.

$sel:elasticGpuSpecifications:RequestLaunchTemplateData', requestLaunchTemplateData_elasticGpuSpecifications - An elastic GPU to associate with the instance.

$sel:capacityReservationSpecification:RequestLaunchTemplateData', requestLaunchTemplateData_capacityReservationSpecification - The Capacity Reservation targeting option. If you do not specify this parameter, the instance's Capacity Reservation preference defaults to open, which enables it to run in any open Capacity Reservation that has matching attributes (instance type, platform, Availability Zone).

$sel:ebsOptimized:RequestLaunchTemplateData', requestLaunchTemplateData_ebsOptimized - Indicates whether the instance is optimized for Amazon EBS I/O. This optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal Amazon EBS I/O performance. This optimization isn't available with all instance types. Additional usage charges apply when using an EBS-optimized instance.

$sel:placement:RequestLaunchTemplateData', requestLaunchTemplateData_placement - The placement for the instance.

$sel:userData:RequestLaunchTemplateData', requestLaunchTemplateData_userData - The user data to make available to the instance. You must provide base64-encoded text. User data is limited to 16 KB. For more information, see Running Commands on Your Linux Instance at Launch (Linux) or Adding User Data (Windows).

If you are creating the launch template for use with Batch, the user data must be provided in the MIME multi-part archive format. For more information, see Amazon EC2 user data in launch templates in the Batch User Guide.

$sel:instanceType:RequestLaunchTemplateData', requestLaunchTemplateData_instanceType - The instance type. For more information, see Instance Types in the Amazon Elastic Compute Cloud User Guide.

$sel:ramDiskId:RequestLaunchTemplateData', requestLaunchTemplateData_ramDiskId - The ID of the RAM disk.

We recommend that you use PV-GRUB instead of kernels and RAM disks. For more information, see User Provided Kernels in the Amazon Elastic Compute Cloud User Guide.

$sel:creditSpecification:RequestLaunchTemplateData', requestLaunchTemplateData_creditSpecification - The credit option for CPU usage of the instance. Valid for T2, T3, or T3a instances only.

$sel:licenseSpecifications:RequestLaunchTemplateData', requestLaunchTemplateData_licenseSpecifications - The license configurations.

$sel:instanceMarketOptions:RequestLaunchTemplateData', requestLaunchTemplateData_instanceMarketOptions - The market (purchasing) option for the instances.

$sel:instanceInitiatedShutdownBehavior:RequestLaunchTemplateData', requestLaunchTemplateData_instanceInitiatedShutdownBehavior - Indicates whether an instance stops or terminates when you initiate shutdown from the instance (using the operating system command for system shutdown).

Default: stop

$sel:elasticInferenceAccelerators:RequestLaunchTemplateData', requestLaunchTemplateData_elasticInferenceAccelerators - The elastic inference accelerator for the instance.

$sel:imageId:RequestLaunchTemplateData', requestLaunchTemplateData_imageId - The ID of the AMI.

$sel:securityGroups:RequestLaunchTemplateData', requestLaunchTemplateData_securityGroups - [EC2-Classic, default VPC] One or more security group names. For a nondefault VPC, you must use security group IDs instead. You cannot specify both a security group ID and security name in the same request.

$sel:iamInstanceProfile:RequestLaunchTemplateData', requestLaunchTemplateData_iamInstanceProfile - The name or Amazon Resource Name (ARN) of an IAM instance profile.

$sel:hibernationOptions:RequestLaunchTemplateData', requestLaunchTemplateData_hibernationOptions - Indicates whether an instance is enabled for hibernation. This parameter is valid only if the instance meets the hibernation prerequisites. For more information, see Hibernate your instance in the Amazon Elastic Compute Cloud User Guide.

$sel:monitoring:RequestLaunchTemplateData', requestLaunchTemplateData_monitoring - The monitoring for the instance.

$sel:blockDeviceMappings:RequestLaunchTemplateData', requestLaunchTemplateData_blockDeviceMappings - The block device mapping.

$sel:enclaveOptions:RequestLaunchTemplateData', requestLaunchTemplateData_enclaveOptions - Indicates whether the instance is enabled for Amazon Web Services Nitro Enclaves. For more information, see What is Amazon Web Services Nitro Enclaves? in the Amazon Web Services Nitro Enclaves User Guide.

You can't enable Amazon Web Services Nitro Enclaves and hibernation on the same instance.

$sel:kernelId:RequestLaunchTemplateData', requestLaunchTemplateData_kernelId - The ID of the kernel.

We recommend that you use PV-GRUB instead of kernels and RAM disks. For more information, see User Provided Kernels in the Amazon Elastic Compute Cloud User Guide.

$sel:cpuOptions:RequestLaunchTemplateData', requestLaunchTemplateData_cpuOptions - The CPU options for the instance. For more information, see Optimizing CPU Options in the Amazon Elastic Compute Cloud User Guide.

$sel:disableApiTermination:RequestLaunchTemplateData', requestLaunchTemplateData_disableApiTermination - If you set this parameter to true, you can't terminate the instance using the Amazon EC2 console, CLI, or API; otherwise, you can. To change this attribute after launch, use ModifyInstanceAttribute. Alternatively, if you set InstanceInitiatedShutdownBehavior to terminate, you can terminate the instance by running the shutdown command from the instance.

$sel:networkInterfaces:RequestLaunchTemplateData', requestLaunchTemplateData_networkInterfaces - One or more network interfaces. If you specify a network interface, you must specify any security groups and subnets as part of the network interface.

$sel:keyName:RequestLaunchTemplateData', requestLaunchTemplateData_keyName - The name of the key pair. You can create a key pair using CreateKeyPair or ImportKeyPair.

If you do not specify a key pair, you can't connect to the instance unless you choose an AMI that is configured to allow users another way to log in.

$sel:metadataOptions:RequestLaunchTemplateData', requestLaunchTemplateData_metadataOptions - The metadata options for the instance. For more information, see Instance metadata and user data in the Amazon Elastic Compute Cloud User Guide.

requestLaunchTemplateData_securityGroupIds :: Lens' RequestLaunchTemplateData (Maybe [Text])

One or more security group IDs. You can create a security group using CreateSecurityGroup. You cannot specify both a security group ID and security name in the same request.

requestLaunchTemplateData_tagSpecifications :: Lens' RequestLaunchTemplateData (Maybe [LaunchTemplateTagSpecificationRequest])

The tags to apply to the resources during launch. You can only tag instances and volumes on launch. The specified tags are applied to all instances or volumes that are created during launch. To tag a resource after it has been created, see CreateTags.

requestLaunchTemplateData_elasticGpuSpecifications :: Lens' RequestLaunchTemplateData (Maybe [ElasticGpuSpecification])

An elastic GPU to associate with the instance.

requestLaunchTemplateData_capacityReservationSpecification :: Lens' RequestLaunchTemplateData (Maybe LaunchTemplateCapacityReservationSpecificationRequest)

The Capacity Reservation targeting option. If you do not specify this parameter, the instance's Capacity Reservation preference defaults to open, which enables it to run in any open Capacity Reservation that has matching attributes (instance type, platform, Availability Zone).

requestLaunchTemplateData_ebsOptimized :: Lens' RequestLaunchTemplateData (Maybe Bool)

Indicates whether the instance is optimized for Amazon EBS I/O. This optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal Amazon EBS I/O performance. This optimization isn't available with all instance types. Additional usage charges apply when using an EBS-optimized instance.

requestLaunchTemplateData_placement :: Lens' RequestLaunchTemplateData (Maybe LaunchTemplatePlacementRequest)

The placement for the instance.

requestLaunchTemplateData_userData :: Lens' RequestLaunchTemplateData (Maybe Text)

The user data to make available to the instance. You must provide base64-encoded text. User data is limited to 16 KB. For more information, see Running Commands on Your Linux Instance at Launch (Linux) or Adding User Data (Windows).

If you are creating the launch template for use with Batch, the user data must be provided in the MIME multi-part archive format. For more information, see Amazon EC2 user data in launch templates in the Batch User Guide.

requestLaunchTemplateData_instanceType :: Lens' RequestLaunchTemplateData (Maybe InstanceType)

The instance type. For more information, see Instance Types in the Amazon Elastic Compute Cloud User Guide.

requestLaunchTemplateData_ramDiskId :: Lens' RequestLaunchTemplateData (Maybe Text)

The ID of the RAM disk.

We recommend that you use PV-GRUB instead of kernels and RAM disks. For more information, see User Provided Kernels in the Amazon Elastic Compute Cloud User Guide.

requestLaunchTemplateData_creditSpecification :: Lens' RequestLaunchTemplateData (Maybe CreditSpecificationRequest)

The credit option for CPU usage of the instance. Valid for T2, T3, or T3a instances only.

requestLaunchTemplateData_licenseSpecifications :: Lens' RequestLaunchTemplateData (Maybe [LaunchTemplateLicenseConfigurationRequest])

The license configurations.

requestLaunchTemplateData_instanceMarketOptions :: Lens' RequestLaunchTemplateData (Maybe LaunchTemplateInstanceMarketOptionsRequest)

The market (purchasing) option for the instances.

requestLaunchTemplateData_instanceInitiatedShutdownBehavior :: Lens' RequestLaunchTemplateData (Maybe ShutdownBehavior)

Indicates whether an instance stops or terminates when you initiate shutdown from the instance (using the operating system command for system shutdown).

Default: stop

requestLaunchTemplateData_elasticInferenceAccelerators :: Lens' RequestLaunchTemplateData (Maybe [LaunchTemplateElasticInferenceAccelerator])

The elastic inference accelerator for the instance.

requestLaunchTemplateData_imageId :: Lens' RequestLaunchTemplateData (Maybe Text)

The ID of the AMI.

requestLaunchTemplateData_securityGroups :: Lens' RequestLaunchTemplateData (Maybe [Text])

EC2-Classic, default VPC

One or more security group names. For a nondefault VPC, you must use security group IDs instead. You cannot specify both a security group ID and security name in the same request.

requestLaunchTemplateData_iamInstanceProfile :: Lens' RequestLaunchTemplateData (Maybe LaunchTemplateIamInstanceProfileSpecificationRequest)

The name or Amazon Resource Name (ARN) of an IAM instance profile.

requestLaunchTemplateData_hibernationOptions :: Lens' RequestLaunchTemplateData (Maybe LaunchTemplateHibernationOptionsRequest)

Indicates whether an instance is enabled for hibernation. This parameter is valid only if the instance meets the hibernation prerequisites. For more information, see Hibernate your instance in the Amazon Elastic Compute Cloud User Guide.

requestLaunchTemplateData_monitoring :: Lens' RequestLaunchTemplateData (Maybe LaunchTemplatesMonitoringRequest)

The monitoring for the instance.

requestLaunchTemplateData_blockDeviceMappings :: Lens' RequestLaunchTemplateData (Maybe [LaunchTemplateBlockDeviceMappingRequest])

The block device mapping.

requestLaunchTemplateData_enclaveOptions :: Lens' RequestLaunchTemplateData (Maybe LaunchTemplateEnclaveOptionsRequest)

Indicates whether the instance is enabled for Amazon Web Services Nitro Enclaves. For more information, see What is Amazon Web Services Nitro Enclaves? in the Amazon Web Services Nitro Enclaves User Guide.

You can't enable Amazon Web Services Nitro Enclaves and hibernation on the same instance.

requestLaunchTemplateData_kernelId :: Lens' RequestLaunchTemplateData (Maybe Text)

The ID of the kernel.

We recommend that you use PV-GRUB instead of kernels and RAM disks. For more information, see User Provided Kernels in the Amazon Elastic Compute Cloud User Guide.

requestLaunchTemplateData_cpuOptions :: Lens' RequestLaunchTemplateData (Maybe LaunchTemplateCpuOptionsRequest)

The CPU options for the instance. For more information, see Optimizing CPU Options in the Amazon Elastic Compute Cloud User Guide.

requestLaunchTemplateData_disableApiTermination :: Lens' RequestLaunchTemplateData (Maybe Bool)

If you set this parameter to true, you can't terminate the instance using the Amazon EC2 console, CLI, or API; otherwise, you can. To change this attribute after launch, use ModifyInstanceAttribute. Alternatively, if you set InstanceInitiatedShutdownBehavior to terminate, you can terminate the instance by running the shutdown command from the instance.

requestLaunchTemplateData_networkInterfaces :: Lens' RequestLaunchTemplateData (Maybe [LaunchTemplateInstanceNetworkInterfaceSpecificationRequest])

One or more network interfaces. If you specify a network interface, you must specify any security groups and subnets as part of the network interface.

requestLaunchTemplateData_keyName :: Lens' RequestLaunchTemplateData (Maybe Text)

The name of the key pair. You can create a key pair using CreateKeyPair or ImportKeyPair.

If you do not specify a key pair, you can't connect to the instance unless you choose an AMI that is configured to allow users another way to log in.

requestLaunchTemplateData_metadataOptions :: Lens' RequestLaunchTemplateData (Maybe LaunchTemplateInstanceMetadataOptionsRequest)

The metadata options for the instance. For more information, see Instance metadata and user data in the Amazon Elastic Compute Cloud User Guide.