Network.AWS.EC2.Types.ReservedInstances

Description

Documentation

data ReservedInstances

Describes a Reserved Instance.

See: newReservedInstances smart constructor.

Constructors

ReservedInstances'

Fields

end :: Maybe ISO8601

The time when the Reserved Instance expires.

instanceType :: Maybe InstanceType

The instance type on which the Reserved Instance can be used.

duration :: Maybe Integer

The duration of the Reserved Instance, in seconds.

reservedInstancesId :: Maybe Text

The ID of the Reserved Instance.

instanceTenancy :: Maybe Tenancy

The tenancy of the instance.

currencyCode :: Maybe CurrencyCodeValues

The currency of the Reserved Instance. It's specified using ISO 4217 standard currency codes. At this time, the only supported currency is USD.

scope :: Maybe Scope

The scope of the Reserved Instance.

availabilityZone :: Maybe Text

The Availability Zone in which the Reserved Instance can be used.

state :: Maybe ReservedInstanceState

The state of the Reserved Instance purchase.

tags :: Maybe [Tag]

Any tags assigned to the resource.

offeringClass :: Maybe OfferingClassType

The offering class of the Reserved Instance.

fixedPrice :: Maybe Double

The purchase price of the Reserved Instance.

usagePrice :: Maybe Double

The usage price of the Reserved Instance, per hour.

offeringType :: Maybe OfferingTypeValues

The Reserved Instance offering type.

start :: Maybe ISO8601

The date and time the Reserved Instance started.

recurringCharges :: Maybe [RecurringCharge]

The recurring charge tag assigned to the resource.

productDescription :: Maybe RIProductDescription

The Reserved Instance product platform description.

instanceCount :: Maybe Int

The number of reservations purchased.

newReservedInstances :: ReservedInstances

Create a value of ReservedInstances with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:end:ReservedInstances', reservedInstances_end - The time when the Reserved Instance expires.

$sel:instanceType:ReservedInstances', reservedInstances_instanceType - The instance type on which the Reserved Instance can be used.

$sel:duration:ReservedInstances', reservedInstances_duration - The duration of the Reserved Instance, in seconds.

$sel:reservedInstancesId:ReservedInstances', reservedInstances_reservedInstancesId - The ID of the Reserved Instance.

$sel:instanceTenancy:ReservedInstances', reservedInstances_instanceTenancy - The tenancy of the instance.

$sel:currencyCode:ReservedInstances', reservedInstances_currencyCode - The currency of the Reserved Instance. It's specified using ISO 4217 standard currency codes. At this time, the only supported currency is USD.

$sel:scope:ReservedInstances', reservedInstances_scope - The scope of the Reserved Instance.

$sel:availabilityZone:ReservedInstances', reservedInstances_availabilityZone - The Availability Zone in which the Reserved Instance can be used.

$sel:state:ReservedInstances', reservedInstances_state - The state of the Reserved Instance purchase.

$sel:tags:ReservedInstances', reservedInstances_tags - Any tags assigned to the resource.

$sel:offeringClass:ReservedInstances', reservedInstances_offeringClass - The offering class of the Reserved Instance.

$sel:fixedPrice:ReservedInstances', reservedInstances_fixedPrice - The purchase price of the Reserved Instance.

$sel:usagePrice:ReservedInstances', reservedInstances_usagePrice - The usage price of the Reserved Instance, per hour.

$sel:offeringType:ReservedInstances', reservedInstances_offeringType - The Reserved Instance offering type.

$sel:start:ReservedInstances', reservedInstances_start - The date and time the Reserved Instance started.

$sel:recurringCharges:ReservedInstances', reservedInstances_recurringCharges - The recurring charge tag assigned to the resource.

$sel:productDescription:ReservedInstances', reservedInstances_productDescription - The Reserved Instance product platform description.

$sel:instanceCount:ReservedInstances', reservedInstances_instanceCount - The number of reservations purchased.

reservedInstances_end :: Lens' ReservedInstances (Maybe UTCTime)

The time when the Reserved Instance expires.

reservedInstances_instanceType :: Lens' ReservedInstances (Maybe InstanceType)

The instance type on which the Reserved Instance can be used.

reservedInstances_duration :: Lens' ReservedInstances (Maybe Integer)

The duration of the Reserved Instance, in seconds.

reservedInstances_reservedInstancesId :: Lens' ReservedInstances (Maybe Text)

The ID of the Reserved Instance.

reservedInstances_instanceTenancy :: Lens' ReservedInstances (Maybe Tenancy)

The tenancy of the instance.

reservedInstances_currencyCode :: Lens' ReservedInstances (Maybe CurrencyCodeValues)

The currency of the Reserved Instance. It's specified using ISO 4217 standard currency codes. At this time, the only supported currency is USD.

reservedInstances_scope :: Lens' ReservedInstances (Maybe Scope)

The scope of the Reserved Instance.

reservedInstances_availabilityZone :: Lens' ReservedInstances (Maybe Text)

The Availability Zone in which the Reserved Instance can be used.

reservedInstances_state :: Lens' ReservedInstances (Maybe ReservedInstanceState)

The state of the Reserved Instance purchase.

reservedInstances_tags :: Lens' ReservedInstances (Maybe [Tag])

Any tags assigned to the resource.

reservedInstances_offeringClass :: Lens' ReservedInstances (Maybe OfferingClassType)

The offering class of the Reserved Instance.

reservedInstances_fixedPrice :: Lens' ReservedInstances (Maybe Double)

The purchase price of the Reserved Instance.

reservedInstances_usagePrice :: Lens' ReservedInstances (Maybe Double)

The usage price of the Reserved Instance, per hour.

reservedInstances_offeringType :: Lens' ReservedInstances (Maybe OfferingTypeValues)

The Reserved Instance offering type.

reservedInstances_start :: Lens' ReservedInstances (Maybe UTCTime)

The date and time the Reserved Instance started.

reservedInstances_recurringCharges :: Lens' ReservedInstances (Maybe [RecurringCharge])

The recurring charge tag assigned to the resource.

reservedInstances_productDescription :: Lens' ReservedInstances (Maybe RIProductDescription)

The Reserved Instance product platform description.

reservedInstances_instanceCount :: Lens' ReservedInstances (Maybe Int)

The number of reservations purchased.