Network.AWS.EC2.Types.ReservedInstancesConfiguration

Description

Documentation

data ReservedInstancesConfiguration

Describes the configuration settings for the modified Reserved Instances.

See: newReservedInstancesConfiguration smart constructor.

Constructors

ReservedInstancesConfiguration'

Fields

platform :: Maybe Text

The network platform of the modified Reserved Instances, which is either EC2-Classic or EC2-VPC.

instanceType :: Maybe InstanceType

The instance type for the modified Reserved Instances.

scope :: Maybe Scope

Whether the Reserved Instance is applied to instances in a Region or instances in a specific Availability Zone.

availabilityZone :: Maybe Text

The Availability Zone for the modified Reserved Instances.

instanceCount :: Maybe Int

The number of modified Reserved Instances.

This is a required field for a request.

newReservedInstancesConfiguration :: ReservedInstancesConfiguration

Create a value of ReservedInstancesConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:platform:ReservedInstancesConfiguration', reservedInstancesConfiguration_platform - The network platform of the modified Reserved Instances, which is either EC2-Classic or EC2-VPC.

$sel:instanceType:ReservedInstancesConfiguration', reservedInstancesConfiguration_instanceType - The instance type for the modified Reserved Instances.

$sel:scope:ReservedInstancesConfiguration', reservedInstancesConfiguration_scope - Whether the Reserved Instance is applied to instances in a Region or instances in a specific Availability Zone.

$sel:availabilityZone:ReservedInstancesConfiguration', reservedInstancesConfiguration_availabilityZone - The Availability Zone for the modified Reserved Instances.

$sel:instanceCount:ReservedInstancesConfiguration', reservedInstancesConfiguration_instanceCount - The number of modified Reserved Instances.

This is a required field for a request.

reservedInstancesConfiguration_platform :: Lens' ReservedInstancesConfiguration (Maybe Text)

The network platform of the modified Reserved Instances, which is either EC2-Classic or EC2-VPC.

reservedInstancesConfiguration_instanceType :: Lens' ReservedInstancesConfiguration (Maybe InstanceType)

The instance type for the modified Reserved Instances.

reservedInstancesConfiguration_scope :: Lens' ReservedInstancesConfiguration (Maybe Scope)

Whether the Reserved Instance is applied to instances in a Region or instances in a specific Availability Zone.

reservedInstancesConfiguration_availabilityZone :: Lens' ReservedInstancesConfiguration (Maybe Text)

The Availability Zone for the modified Reserved Instances.

reservedInstancesConfiguration_instanceCount :: Lens' ReservedInstancesConfiguration (Maybe Int)

The number of modified Reserved Instances.

This is a required field for a request.