Network.AWS.EC2.Types.ReservedInstancesOffering

Description

Documentation

data ReservedInstancesOffering

Describes a Reserved Instance offering.

See: newReservedInstancesOffering smart constructor.

Constructors

ReservedInstancesOffering'

Fields

instanceType :: Maybe InstanceType

The instance type on which the Reserved Instance can be used.

duration :: Maybe Integer

The duration of the Reserved Instance, in seconds.

reservedInstancesOfferingId :: Maybe Text

The ID of the Reserved Instance offering. This is the offering ID used in GetReservedInstancesExchangeQuote to confirm that an exchange can be made.

instanceTenancy :: Maybe Tenancy

The tenancy of the instance.

currencyCode :: Maybe CurrencyCodeValues

The currency of the Reserved Instance offering you are purchasing. It's specified using ISO 4217 standard currency codes. At this time, the only supported currency is USD.

scope :: Maybe Scope

Whether the Reserved Instance is applied to instances in a Region or an Availability Zone.

availabilityZone :: Maybe Text

The Availability Zone in which the Reserved Instance can be used.

marketplace :: Maybe Bool

Indicates whether the offering is available through the Reserved Instance Marketplace (resale) or Amazon Web Services. If it's a Reserved Instance Marketplace offering, this is true.

offeringClass :: Maybe OfferingClassType

If convertible it can be exchanged for Reserved Instances of the same or higher monetary value, with different configurations. If standard, it is not possible to perform an exchange.

fixedPrice :: Maybe Double

The purchase price of the Reserved Instance.

usagePrice :: Maybe Double

The usage price of the Reserved Instance, per hour.

offeringType :: Maybe OfferingTypeValues

The Reserved Instance offering type.

recurringCharges :: Maybe [RecurringCharge]

The recurring charge tag assigned to the resource.

productDescription :: Maybe RIProductDescription

The Reserved Instance product platform description.

pricingDetails :: Maybe [PricingDetail]

The pricing details of the Reserved Instance offering.

newReservedInstancesOffering :: ReservedInstancesOffering

Create a value of ReservedInstancesOffering with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceType:ReservedInstancesOffering', reservedInstancesOffering_instanceType - The instance type on which the Reserved Instance can be used.

$sel:duration:ReservedInstancesOffering', reservedInstancesOffering_duration - The duration of the Reserved Instance, in seconds.

$sel:reservedInstancesOfferingId:ReservedInstancesOffering', reservedInstancesOffering_reservedInstancesOfferingId - The ID of the Reserved Instance offering. This is the offering ID used in GetReservedInstancesExchangeQuote to confirm that an exchange can be made.

$sel:instanceTenancy:ReservedInstancesOffering', reservedInstancesOffering_instanceTenancy - The tenancy of the instance.

$sel:currencyCode:ReservedInstancesOffering', reservedInstancesOffering_currencyCode - The currency of the Reserved Instance offering you are purchasing. It's specified using ISO 4217 standard currency codes. At this time, the only supported currency is USD.

$sel:scope:ReservedInstancesOffering', reservedInstancesOffering_scope - Whether the Reserved Instance is applied to instances in a Region or an Availability Zone.

$sel:availabilityZone:ReservedInstancesOffering', reservedInstancesOffering_availabilityZone - The Availability Zone in which the Reserved Instance can be used.

$sel:marketplace:ReservedInstancesOffering', reservedInstancesOffering_marketplace - Indicates whether the offering is available through the Reserved Instance Marketplace (resale) or Amazon Web Services. If it's a Reserved Instance Marketplace offering, this is true.

$sel:offeringClass:ReservedInstancesOffering', reservedInstancesOffering_offeringClass - If convertible it can be exchanged for Reserved Instances of the same or higher monetary value, with different configurations. If standard, it is not possible to perform an exchange.

$sel:fixedPrice:ReservedInstancesOffering', reservedInstancesOffering_fixedPrice - The purchase price of the Reserved Instance.

$sel:usagePrice:ReservedInstancesOffering', reservedInstancesOffering_usagePrice - The usage price of the Reserved Instance, per hour.

$sel:offeringType:ReservedInstancesOffering', reservedInstancesOffering_offeringType - The Reserved Instance offering type.

$sel:recurringCharges:ReservedInstancesOffering', reservedInstancesOffering_recurringCharges - The recurring charge tag assigned to the resource.

$sel:productDescription:ReservedInstancesOffering', reservedInstancesOffering_productDescription - The Reserved Instance product platform description.

$sel:pricingDetails:ReservedInstancesOffering', reservedInstancesOffering_pricingDetails - The pricing details of the Reserved Instance offering.

reservedInstancesOffering_instanceType :: Lens' ReservedInstancesOffering (Maybe InstanceType)

The instance type on which the Reserved Instance can be used.

reservedInstancesOffering_duration :: Lens' ReservedInstancesOffering (Maybe Integer)

The duration of the Reserved Instance, in seconds.

reservedInstancesOffering_reservedInstancesOfferingId :: Lens' ReservedInstancesOffering (Maybe Text)

The ID of the Reserved Instance offering. This is the offering ID used in GetReservedInstancesExchangeQuote to confirm that an exchange can be made.

reservedInstancesOffering_instanceTenancy :: Lens' ReservedInstancesOffering (Maybe Tenancy)

The tenancy of the instance.

reservedInstancesOffering_currencyCode :: Lens' ReservedInstancesOffering (Maybe CurrencyCodeValues)

The currency of the Reserved Instance offering you are purchasing. It's specified using ISO 4217 standard currency codes. At this time, the only supported currency is USD.

reservedInstancesOffering_scope :: Lens' ReservedInstancesOffering (Maybe Scope)

Whether the Reserved Instance is applied to instances in a Region or an Availability Zone.

reservedInstancesOffering_availabilityZone :: Lens' ReservedInstancesOffering (Maybe Text)

The Availability Zone in which the Reserved Instance can be used.

reservedInstancesOffering_marketplace :: Lens' ReservedInstancesOffering (Maybe Bool)

Indicates whether the offering is available through the Reserved Instance Marketplace (resale) or Amazon Web Services. If it's a Reserved Instance Marketplace offering, this is true.

reservedInstancesOffering_offeringClass :: Lens' ReservedInstancesOffering (Maybe OfferingClassType)

If convertible it can be exchanged for Reserved Instances of the same or higher monetary value, with different configurations. If standard, it is not possible to perform an exchange.

reservedInstancesOffering_fixedPrice :: Lens' ReservedInstancesOffering (Maybe Double)

The purchase price of the Reserved Instance.

reservedInstancesOffering_usagePrice :: Lens' ReservedInstancesOffering (Maybe Double)

The usage price of the Reserved Instance, per hour.

reservedInstancesOffering_offeringType :: Lens' ReservedInstancesOffering (Maybe OfferingTypeValues)

The Reserved Instance offering type.

reservedInstancesOffering_recurringCharges :: Lens' ReservedInstancesOffering (Maybe [RecurringCharge])

The recurring charge tag assigned to the resource.

reservedInstancesOffering_productDescription :: Lens' ReservedInstancesOffering (Maybe RIProductDescription)

The Reserved Instance product platform description.

reservedInstancesOffering_pricingDetails :: Lens' ReservedInstancesOffering (Maybe [PricingDetail])

The pricing details of the Reserved Instance offering.