Network.AWS.EC2.Types.ResponseLaunchTemplateData

Description

Documentation

data ResponseLaunchTemplateData

The information for a launch template.

See: newResponseLaunchTemplateData smart constructor.

Constructors

ResponseLaunchTemplateData'

Fields

securityGroupIds :: Maybe [Text]

The security group IDs.

tagSpecifications :: Maybe [LaunchTemplateTagSpecification]

The tags.

elasticGpuSpecifications :: Maybe [ElasticGpuSpecificationResponse]

The elastic GPU specification.

capacityReservationSpecification :: Maybe LaunchTemplateCapacityReservationSpecificationResponse

Information about the Capacity Reservation targeting option.

ebsOptimized :: Maybe Bool

Indicates whether the instance is optimized for Amazon EBS I/O.

placement :: Maybe LaunchTemplatePlacement

The placement of the instance.

userData :: Maybe Text

The user data for the instance.

instanceType :: Maybe InstanceType

The instance type.

ramDiskId :: Maybe Text

The ID of the RAM disk, if applicable.

creditSpecification :: Maybe CreditSpecification

The credit option for CPU usage of the instance.

licenseSpecifications :: Maybe [LaunchTemplateLicenseConfiguration]

The license configurations.

instanceMarketOptions :: Maybe LaunchTemplateInstanceMarketOptions

The market (purchasing) option for the instances.

instanceInitiatedShutdownBehavior :: Maybe ShutdownBehavior

Indicates whether an instance stops or terminates when you initiate shutdown from the instance (using the operating system command for system shutdown).

elasticInferenceAccelerators :: Maybe [LaunchTemplateElasticInferenceAcceleratorResponse]

The elastic inference accelerator for the instance.

imageId :: Maybe Text

The ID of the AMI that was used to launch the instance.

securityGroups :: Maybe [Text]

The security group names.

iamInstanceProfile :: Maybe LaunchTemplateIamInstanceProfileSpecification

The IAM instance profile.

hibernationOptions :: Maybe LaunchTemplateHibernationOptions

Indicates whether an instance is configured for hibernation. For more information, see Hibernate your instance in the Amazon Elastic Compute Cloud User Guide.

monitoring :: Maybe LaunchTemplatesMonitoring

The monitoring for the instance.

blockDeviceMappings :: Maybe [LaunchTemplateBlockDeviceMapping]

The block device mappings.

enclaveOptions :: Maybe LaunchTemplateEnclaveOptions

Indicates whether the instance is enabled for Amazon Web Services Nitro Enclaves.

kernelId :: Maybe Text

The ID of the kernel, if applicable.

cpuOptions :: Maybe LaunchTemplateCpuOptions

The CPU options for the instance. For more information, see Optimizing CPU options in the Amazon Elastic Compute Cloud User Guide.

disableApiTermination :: Maybe Bool

If set to true, indicates that the instance cannot be terminated using the Amazon EC2 console, command line tool, or API.

networkInterfaces :: Maybe [LaunchTemplateInstanceNetworkInterfaceSpecification]

The network interfaces.

keyName :: Maybe Text

The name of the key pair.

metadataOptions :: Maybe LaunchTemplateInstanceMetadataOptions

The metadata options for the instance. For more information, see Instance metadata and user data in the Amazon Elastic Compute Cloud User Guide.

newResponseLaunchTemplateData :: ResponseLaunchTemplateData

Create a value of ResponseLaunchTemplateData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:securityGroupIds:ResponseLaunchTemplateData', responseLaunchTemplateData_securityGroupIds - The security group IDs.

$sel:tagSpecifications:ResponseLaunchTemplateData', responseLaunchTemplateData_tagSpecifications - The tags.

$sel:elasticGpuSpecifications:ResponseLaunchTemplateData', responseLaunchTemplateData_elasticGpuSpecifications - The elastic GPU specification.

$sel:capacityReservationSpecification:ResponseLaunchTemplateData', responseLaunchTemplateData_capacityReservationSpecification - Information about the Capacity Reservation targeting option.

$sel:ebsOptimized:ResponseLaunchTemplateData', responseLaunchTemplateData_ebsOptimized - Indicates whether the instance is optimized for Amazon EBS I/O.

$sel:placement:ResponseLaunchTemplateData', responseLaunchTemplateData_placement - The placement of the instance.

$sel:userData:ResponseLaunchTemplateData', responseLaunchTemplateData_userData - The user data for the instance.

$sel:instanceType:ResponseLaunchTemplateData', responseLaunchTemplateData_instanceType - The instance type.

$sel:ramDiskId:ResponseLaunchTemplateData', responseLaunchTemplateData_ramDiskId - The ID of the RAM disk, if applicable.

$sel:creditSpecification:ResponseLaunchTemplateData', responseLaunchTemplateData_creditSpecification - The credit option for CPU usage of the instance.

$sel:licenseSpecifications:ResponseLaunchTemplateData', responseLaunchTemplateData_licenseSpecifications - The license configurations.

$sel:instanceMarketOptions:ResponseLaunchTemplateData', responseLaunchTemplateData_instanceMarketOptions - The market (purchasing) option for the instances.

$sel:instanceInitiatedShutdownBehavior:ResponseLaunchTemplateData', responseLaunchTemplateData_instanceInitiatedShutdownBehavior - Indicates whether an instance stops or terminates when you initiate shutdown from the instance (using the operating system command for system shutdown).

$sel:elasticInferenceAccelerators:ResponseLaunchTemplateData', responseLaunchTemplateData_elasticInferenceAccelerators - The elastic inference accelerator for the instance.

$sel:imageId:ResponseLaunchTemplateData', responseLaunchTemplateData_imageId - The ID of the AMI that was used to launch the instance.

$sel:securityGroups:ResponseLaunchTemplateData', responseLaunchTemplateData_securityGroups - The security group names.

$sel:iamInstanceProfile:ResponseLaunchTemplateData', responseLaunchTemplateData_iamInstanceProfile - The IAM instance profile.

$sel:hibernationOptions:ResponseLaunchTemplateData', responseLaunchTemplateData_hibernationOptions - Indicates whether an instance is configured for hibernation. For more information, see Hibernate your instance in the Amazon Elastic Compute Cloud User Guide.

$sel:monitoring:ResponseLaunchTemplateData', responseLaunchTemplateData_monitoring - The monitoring for the instance.

$sel:blockDeviceMappings:ResponseLaunchTemplateData', responseLaunchTemplateData_blockDeviceMappings - The block device mappings.

$sel:enclaveOptions:ResponseLaunchTemplateData', responseLaunchTemplateData_enclaveOptions - Indicates whether the instance is enabled for Amazon Web Services Nitro Enclaves.

$sel:kernelId:ResponseLaunchTemplateData', responseLaunchTemplateData_kernelId - The ID of the kernel, if applicable.

$sel:cpuOptions:ResponseLaunchTemplateData', responseLaunchTemplateData_cpuOptions - The CPU options for the instance. For more information, see Optimizing CPU options in the Amazon Elastic Compute Cloud User Guide.

$sel:disableApiTermination:ResponseLaunchTemplateData', responseLaunchTemplateData_disableApiTermination - If set to true, indicates that the instance cannot be terminated using the Amazon EC2 console, command line tool, or API.

$sel:networkInterfaces:ResponseLaunchTemplateData', responseLaunchTemplateData_networkInterfaces - The network interfaces.

$sel:keyName:ResponseLaunchTemplateData', responseLaunchTemplateData_keyName - The name of the key pair.

$sel:metadataOptions:ResponseLaunchTemplateData', responseLaunchTemplateData_metadataOptions - The metadata options for the instance. For more information, see Instance metadata and user data in the Amazon Elastic Compute Cloud User Guide.

responseLaunchTemplateData_securityGroupIds :: Lens' ResponseLaunchTemplateData (Maybe [Text])

The security group IDs.

responseLaunchTemplateData_tagSpecifications :: Lens' ResponseLaunchTemplateData (Maybe [LaunchTemplateTagSpecification])

The tags.

responseLaunchTemplateData_elasticGpuSpecifications :: Lens' ResponseLaunchTemplateData (Maybe [ElasticGpuSpecificationResponse])

The elastic GPU specification.

responseLaunchTemplateData_capacityReservationSpecification :: Lens' ResponseLaunchTemplateData (Maybe LaunchTemplateCapacityReservationSpecificationResponse)

Information about the Capacity Reservation targeting option.

responseLaunchTemplateData_ebsOptimized :: Lens' ResponseLaunchTemplateData (Maybe Bool)

Indicates whether the instance is optimized for Amazon EBS I/O.

responseLaunchTemplateData_placement :: Lens' ResponseLaunchTemplateData (Maybe LaunchTemplatePlacement)

The placement of the instance.

responseLaunchTemplateData_userData :: Lens' ResponseLaunchTemplateData (Maybe Text)

The user data for the instance.

responseLaunchTemplateData_instanceType :: Lens' ResponseLaunchTemplateData (Maybe InstanceType)

The instance type.

responseLaunchTemplateData_ramDiskId :: Lens' ResponseLaunchTemplateData (Maybe Text)

The ID of the RAM disk, if applicable.

responseLaunchTemplateData_creditSpecification :: Lens' ResponseLaunchTemplateData (Maybe CreditSpecification)

The credit option for CPU usage of the instance.

responseLaunchTemplateData_licenseSpecifications :: Lens' ResponseLaunchTemplateData (Maybe [LaunchTemplateLicenseConfiguration])

The license configurations.

responseLaunchTemplateData_instanceMarketOptions :: Lens' ResponseLaunchTemplateData (Maybe LaunchTemplateInstanceMarketOptions)

The market (purchasing) option for the instances.

responseLaunchTemplateData_instanceInitiatedShutdownBehavior :: Lens' ResponseLaunchTemplateData (Maybe ShutdownBehavior)

Indicates whether an instance stops or terminates when you initiate shutdown from the instance (using the operating system command for system shutdown).

responseLaunchTemplateData_elasticInferenceAccelerators :: Lens' ResponseLaunchTemplateData (Maybe [LaunchTemplateElasticInferenceAcceleratorResponse])

The elastic inference accelerator for the instance.

responseLaunchTemplateData_imageId :: Lens' ResponseLaunchTemplateData (Maybe Text)

The ID of the AMI that was used to launch the instance.

responseLaunchTemplateData_securityGroups :: Lens' ResponseLaunchTemplateData (Maybe [Text])

The security group names.

responseLaunchTemplateData_iamInstanceProfile :: Lens' ResponseLaunchTemplateData (Maybe LaunchTemplateIamInstanceProfileSpecification)

The IAM instance profile.

responseLaunchTemplateData_hibernationOptions :: Lens' ResponseLaunchTemplateData (Maybe LaunchTemplateHibernationOptions)

Indicates whether an instance is configured for hibernation. For more information, see Hibernate your instance in the Amazon Elastic Compute Cloud User Guide.

responseLaunchTemplateData_monitoring :: Lens' ResponseLaunchTemplateData (Maybe LaunchTemplatesMonitoring)

The monitoring for the instance.

responseLaunchTemplateData_blockDeviceMappings :: Lens' ResponseLaunchTemplateData (Maybe [LaunchTemplateBlockDeviceMapping])

The block device mappings.

responseLaunchTemplateData_enclaveOptions :: Lens' ResponseLaunchTemplateData (Maybe LaunchTemplateEnclaveOptions)

Indicates whether the instance is enabled for Amazon Web Services Nitro Enclaves.

responseLaunchTemplateData_kernelId :: Lens' ResponseLaunchTemplateData (Maybe Text)

The ID of the kernel, if applicable.

responseLaunchTemplateData_cpuOptions :: Lens' ResponseLaunchTemplateData (Maybe LaunchTemplateCpuOptions)

The CPU options for the instance. For more information, see Optimizing CPU options in the Amazon Elastic Compute Cloud User Guide.

responseLaunchTemplateData_disableApiTermination :: Lens' ResponseLaunchTemplateData (Maybe Bool)

If set to true, indicates that the instance cannot be terminated using the Amazon EC2 console, command line tool, or API.

responseLaunchTemplateData_networkInterfaces :: Lens' ResponseLaunchTemplateData (Maybe [LaunchTemplateInstanceNetworkInterfaceSpecification])

The network interfaces.

responseLaunchTemplateData_keyName :: Lens' ResponseLaunchTemplateData (Maybe Text)

The name of the key pair.

responseLaunchTemplateData_metadataOptions :: Lens' ResponseLaunchTemplateData (Maybe LaunchTemplateInstanceMetadataOptions)

The metadata options for the instance. For more information, see Instance metadata and user data in the Amazon Elastic Compute Cloud User Guide.