Network.AWS.EC2.Types.RouteTable

Description

Documentation

data RouteTable

Describes a route table.

See: newRouteTable smart constructor.

Constructors

RouteTable'

Fields

ownerId :: Maybe Text

The ID of the Amazon Web Services account that owns the route table.

routes :: Maybe [Route]

The routes in the route table.

routeTableId :: Maybe Text

The ID of the route table.

tags :: Maybe [Tag]

Any tags assigned to the route table.

propagatingVgws :: Maybe [PropagatingVgw]

Any virtual private gateway (VGW) propagating routes.

vpcId :: Maybe Text

The ID of the VPC.

associations :: Maybe [RouteTableAssociation]

The associations between the route table and one or more subnets or a gateway.

newRouteTable :: RouteTable

Create a value of RouteTable with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ownerId:RouteTable', routeTable_ownerId - The ID of the Amazon Web Services account that owns the route table.

$sel:routes:RouteTable', routeTable_routes - The routes in the route table.

$sel:routeTableId:RouteTable', routeTable_routeTableId - The ID of the route table.

$sel:tags:RouteTable', routeTable_tags - Any tags assigned to the route table.

$sel:propagatingVgws:RouteTable', routeTable_propagatingVgws - Any virtual private gateway (VGW) propagating routes.

$sel:vpcId:RouteTable', routeTable_vpcId - The ID of the VPC.

$sel:associations:RouteTable', routeTable_associations - The associations between the route table and one or more subnets or a gateway.

routeTable_ownerId :: Lens' RouteTable (Maybe Text)

The ID of the Amazon Web Services account that owns the route table.

routeTable_routes :: Lens' RouteTable (Maybe [Route])

The routes in the route table.

routeTable_routeTableId :: Lens' RouteTable (Maybe Text)

The ID of the route table.

routeTable_tags :: Lens' RouteTable (Maybe [Tag])

Any tags assigned to the route table.

routeTable_propagatingVgws :: Lens' RouteTable (Maybe [PropagatingVgw])

Any virtual private gateway (VGW) propagating routes.

routeTable_vpcId :: Lens' RouteTable (Maybe Text)

The ID of the VPC.

routeTable_associations :: Lens' RouteTable (Maybe [RouteTableAssociation])

The associations between the route table and one or more subnets or a gateway.