Network.AWS.EC2.Types.RouteTableAssociation

Description

Documentation

data RouteTableAssociation

Describes an association between a route table and a subnet or gateway.

See: newRouteTableAssociation smart constructor.

Constructors

RouteTableAssociation'

Fields

associationState :: Maybe RouteTableAssociationState

The state of the association.

main :: Maybe Bool

Indicates whether this is the main route table.

routeTableId :: Maybe Text

The ID of the route table.

subnetId :: Maybe Text

The ID of the subnet. A subnet ID is not returned for an implicit association.

routeTableAssociationId :: Maybe Text

The ID of the association.

gatewayId :: Maybe Text

The ID of the internet gateway or virtual private gateway.

newRouteTableAssociation :: RouteTableAssociation

Create a value of RouteTableAssociation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:associationState:RouteTableAssociation', routeTableAssociation_associationState - The state of the association.

$sel:main:RouteTableAssociation', routeTableAssociation_main - Indicates whether this is the main route table.

$sel:routeTableId:RouteTableAssociation', routeTableAssociation_routeTableId - The ID of the route table.

$sel:subnetId:RouteTableAssociation', routeTableAssociation_subnetId - The ID of the subnet. A subnet ID is not returned for an implicit association.

$sel:routeTableAssociationId:RouteTableAssociation', routeTableAssociation_routeTableAssociationId - The ID of the association.

$sel:gatewayId:RouteTableAssociation', routeTableAssociation_gatewayId - The ID of the internet gateway or virtual private gateway.

routeTableAssociation_associationState :: Lens' RouteTableAssociation (Maybe RouteTableAssociationState)

The state of the association.

routeTableAssociation_main :: Lens' RouteTableAssociation (Maybe Bool)

Indicates whether this is the main route table.

routeTableAssociation_routeTableId :: Lens' RouteTableAssociation (Maybe Text)

The ID of the route table.

routeTableAssociation_subnetId :: Lens' RouteTableAssociation (Maybe Text)

The ID of the subnet. A subnet ID is not returned for an implicit association.

routeTableAssociation_routeTableAssociationId :: Lens' RouteTableAssociation (Maybe Text)

The ID of the association.

routeTableAssociation_gatewayId :: Lens' RouteTableAssociation (Maybe Text)

The ID of the internet gateway or virtual private gateway.