Network.AWS.EC2.Types.S3Storage

Description

Documentation

data S3Storage

Describes the storage parameters for Amazon S3 and Amazon S3 buckets for an instance store-backed AMI.

See: newS3Storage smart constructor.

Constructors

S3Storage'

Fields

prefix :: Maybe Text

The beginning of the file name of the AMI.

uploadPolicy :: Maybe Base64

An Amazon S3 upload policy that gives Amazon EC2 permission to upload items into Amazon S3 on your behalf.

bucket :: Maybe Text

The bucket in which to store the AMI. You can specify a bucket that you already own or a new bucket that Amazon EC2 creates on your behalf. If you specify a bucket that belongs to someone else, Amazon EC2 returns an error.

uploadPolicySignature :: Maybe Text

The signature of the JSON document.

aWSAccessKeyId :: Maybe Text

The access key ID of the owner of the bucket. Before you specify a value for your access key ID, review and follow the guidance in Best Practices for Managing Amazon Web Services Access Keys.

newS3Storage :: S3Storage

Create a value of S3Storage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:prefix:S3Storage', s3Storage_prefix - The beginning of the file name of the AMI.

$sel:uploadPolicy:S3Storage', s3Storage_uploadPolicy - An Amazon S3 upload policy that gives Amazon EC2 permission to upload items into Amazon S3 on your behalf.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:bucket:S3Storage', s3Storage_bucket - The bucket in which to store the AMI. You can specify a bucket that you already own or a new bucket that Amazon EC2 creates on your behalf. If you specify a bucket that belongs to someone else, Amazon EC2 returns an error.

$sel:uploadPolicySignature:S3Storage', s3Storage_uploadPolicySignature - The signature of the JSON document.

$sel:aWSAccessKeyId:S3Storage', s3Storage_aWSAccessKeyId - The access key ID of the owner of the bucket. Before you specify a value for your access key ID, review and follow the guidance in Best Practices for Managing Amazon Web Services Access Keys.

s3Storage_prefix :: Lens' S3Storage (Maybe Text)

The beginning of the file name of the AMI.

s3Storage_uploadPolicy :: Lens' S3Storage (Maybe ByteString)

An Amazon S3 upload policy that gives Amazon EC2 permission to upload items into Amazon S3 on your behalf.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

s3Storage_bucket :: Lens' S3Storage (Maybe Text)

The bucket in which to store the AMI. You can specify a bucket that you already own or a new bucket that Amazon EC2 creates on your behalf. If you specify a bucket that belongs to someone else, Amazon EC2 returns an error.

s3Storage_uploadPolicySignature :: Lens' S3Storage (Maybe Text)

The signature of the JSON document.

s3Storage_aWSAccessKeyId :: Lens' S3Storage (Maybe Text)

The access key ID of the owner of the bucket. Before you specify a value for your access key ID, review and follow the guidance in Best Practices for Managing Amazon Web Services Access Keys.