Network.AWS.EC2.Types.ScheduledInstance

Description

Documentation

data ScheduledInstance

Describes a Scheduled Instance.

See: newScheduledInstance smart constructor.

Constructors

ScheduledInstance'

Fields

platform :: Maybe Text

The platform (Linux/UNIX or Windows).

networkPlatform :: Maybe Text

The network platform (EC2-Classic or EC2-VPC).

instanceType :: Maybe Text

The instance type.

slotDurationInHours :: Maybe Int

The number of hours in the schedule.

createDate :: Maybe ISO8601

The date when the Scheduled Instance was purchased.

scheduledInstanceId :: Maybe Text

The Scheduled Instance ID.

availabilityZone :: Maybe Text

The Availability Zone.

previousSlotEndTime :: Maybe ISO8601

The time that the previous schedule ended or will end.

recurrence :: Maybe ScheduledInstanceRecurrence

The schedule recurrence.

totalScheduledInstanceHours :: Maybe Int

The total number of hours for a single instance for the entire term.

nextSlotStartTime :: Maybe ISO8601

The time for the next schedule to start.

hourlyPrice :: Maybe Text

The hourly price for a single instance.

termEndDate :: Maybe ISO8601

The end date for the Scheduled Instance.

termStartDate :: Maybe ISO8601

The start date for the Scheduled Instance.

instanceCount :: Maybe Int

The number of instances.

newScheduledInstance :: ScheduledInstance

Create a value of ScheduledInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:platform:ScheduledInstance', scheduledInstance_platform - The platform (Linux/UNIX or Windows).

$sel:networkPlatform:ScheduledInstance', scheduledInstance_networkPlatform - The network platform (EC2-Classic or EC2-VPC).

$sel:instanceType:ScheduledInstance', scheduledInstance_instanceType - The instance type.

$sel:slotDurationInHours:ScheduledInstance', scheduledInstance_slotDurationInHours - The number of hours in the schedule.

$sel:createDate:ScheduledInstance', scheduledInstance_createDate - The date when the Scheduled Instance was purchased.

$sel:scheduledInstanceId:ScheduledInstance', scheduledInstance_scheduledInstanceId - The Scheduled Instance ID.

$sel:availabilityZone:ScheduledInstance', scheduledInstance_availabilityZone - The Availability Zone.

$sel:previousSlotEndTime:ScheduledInstance', scheduledInstance_previousSlotEndTime - The time that the previous schedule ended or will end.

$sel:recurrence:ScheduledInstance', scheduledInstance_recurrence - The schedule recurrence.

$sel:totalScheduledInstanceHours:ScheduledInstance', scheduledInstance_totalScheduledInstanceHours - The total number of hours for a single instance for the entire term.

$sel:nextSlotStartTime:ScheduledInstance', scheduledInstance_nextSlotStartTime - The time for the next schedule to start.

$sel:hourlyPrice:ScheduledInstance', scheduledInstance_hourlyPrice - The hourly price for a single instance.

$sel:termEndDate:ScheduledInstance', scheduledInstance_termEndDate - The end date for the Scheduled Instance.

$sel:termStartDate:ScheduledInstance', scheduledInstance_termStartDate - The start date for the Scheduled Instance.

$sel:instanceCount:ScheduledInstance', scheduledInstance_instanceCount - The number of instances.

scheduledInstance_platform :: Lens' ScheduledInstance (Maybe Text)

The platform (Linux/UNIX or Windows).

scheduledInstance_networkPlatform :: Lens' ScheduledInstance (Maybe Text)

The network platform (EC2-Classic or EC2-VPC).

scheduledInstance_instanceType :: Lens' ScheduledInstance (Maybe Text)

The instance type.

scheduledInstance_slotDurationInHours :: Lens' ScheduledInstance (Maybe Int)

The number of hours in the schedule.

scheduledInstance_createDate :: Lens' ScheduledInstance (Maybe UTCTime)

The date when the Scheduled Instance was purchased.

scheduledInstance_scheduledInstanceId :: Lens' ScheduledInstance (Maybe Text)

The Scheduled Instance ID.

scheduledInstance_availabilityZone :: Lens' ScheduledInstance (Maybe Text)

The Availability Zone.

scheduledInstance_previousSlotEndTime :: Lens' ScheduledInstance (Maybe UTCTime)

The time that the previous schedule ended or will end.

scheduledInstance_recurrence :: Lens' ScheduledInstance (Maybe ScheduledInstanceRecurrence)

The schedule recurrence.

scheduledInstance_totalScheduledInstanceHours :: Lens' ScheduledInstance (Maybe Int)

The total number of hours for a single instance for the entire term.

scheduledInstance_nextSlotStartTime :: Lens' ScheduledInstance (Maybe UTCTime)

The time for the next schedule to start.

scheduledInstance_hourlyPrice :: Lens' ScheduledInstance (Maybe Text)

The hourly price for a single instance.

scheduledInstance_termEndDate :: Lens' ScheduledInstance (Maybe UTCTime)

The end date for the Scheduled Instance.

scheduledInstance_termStartDate :: Lens' ScheduledInstance (Maybe UTCTime)

The start date for the Scheduled Instance.

scheduledInstance_instanceCount :: Lens' ScheduledInstance (Maybe Int)

The number of instances.