Network.AWS.EC2.Types.ScheduledInstanceAvailability

Description

Documentation

data ScheduledInstanceAvailability

Describes a schedule that is available for your Scheduled Instances.

See: newScheduledInstanceAvailability smart constructor.

Constructors

ScheduledInstanceAvailability'

Fields

platform :: Maybe Text

The platform (Linux/UNIX or Windows).

networkPlatform :: Maybe Text

The network platform (EC2-Classic or EC2-VPC).

instanceType :: Maybe Text

The instance type. You can specify one of the C3, C4, M4, or R3 instance types.

slotDurationInHours :: Maybe Int

The number of hours in the schedule.

availableInstanceCount :: Maybe Int

The number of available instances.

minTermDurationInDays :: Maybe Int

The minimum term. The only possible value is 365 days.

availabilityZone :: Maybe Text

The Availability Zone.

maxTermDurationInDays :: Maybe Int

The maximum term. The only possible value is 365 days.

recurrence :: Maybe ScheduledInstanceRecurrence

The schedule recurrence.

totalScheduledInstanceHours :: Maybe Int

The total number of hours for a single instance for the entire term.

firstSlotStartTime :: Maybe ISO8601

The time period for the first schedule to start.

purchaseToken :: Maybe Text

The purchase token. This token expires in two hours.

hourlyPrice :: Maybe Text

The hourly price for a single instance.

newScheduledInstanceAvailability :: ScheduledInstanceAvailability

Create a value of ScheduledInstanceAvailability with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:platform:ScheduledInstanceAvailability', scheduledInstanceAvailability_platform - The platform (Linux/UNIX or Windows).

$sel:networkPlatform:ScheduledInstanceAvailability', scheduledInstanceAvailability_networkPlatform - The network platform (EC2-Classic or EC2-VPC).

$sel:instanceType:ScheduledInstanceAvailability', scheduledInstanceAvailability_instanceType - The instance type. You can specify one of the C3, C4, M4, or R3 instance types.

$sel:slotDurationInHours:ScheduledInstanceAvailability', scheduledInstanceAvailability_slotDurationInHours - The number of hours in the schedule.

$sel:availableInstanceCount:ScheduledInstanceAvailability', scheduledInstanceAvailability_availableInstanceCount - The number of available instances.

$sel:minTermDurationInDays:ScheduledInstanceAvailability', scheduledInstanceAvailability_minTermDurationInDays - The minimum term. The only possible value is 365 days.

$sel:availabilityZone:ScheduledInstanceAvailability', scheduledInstanceAvailability_availabilityZone - The Availability Zone.

$sel:maxTermDurationInDays:ScheduledInstanceAvailability', scheduledInstanceAvailability_maxTermDurationInDays - The maximum term. The only possible value is 365 days.

$sel:recurrence:ScheduledInstanceAvailability', scheduledInstanceAvailability_recurrence - The schedule recurrence.

$sel:totalScheduledInstanceHours:ScheduledInstanceAvailability', scheduledInstanceAvailability_totalScheduledInstanceHours - The total number of hours for a single instance for the entire term.

$sel:firstSlotStartTime:ScheduledInstanceAvailability', scheduledInstanceAvailability_firstSlotStartTime - The time period for the first schedule to start.

$sel:purchaseToken:ScheduledInstanceAvailability', scheduledInstanceAvailability_purchaseToken - The purchase token. This token expires in two hours.

$sel:hourlyPrice:ScheduledInstanceAvailability', scheduledInstanceAvailability_hourlyPrice - The hourly price for a single instance.

scheduledInstanceAvailability_platform :: Lens' ScheduledInstanceAvailability (Maybe Text)

The platform (Linux/UNIX or Windows).

scheduledInstanceAvailability_networkPlatform :: Lens' ScheduledInstanceAvailability (Maybe Text)

The network platform (EC2-Classic or EC2-VPC).

scheduledInstanceAvailability_instanceType :: Lens' ScheduledInstanceAvailability (Maybe Text)

The instance type. You can specify one of the C3, C4, M4, or R3 instance types.

scheduledInstanceAvailability_slotDurationInHours :: Lens' ScheduledInstanceAvailability (Maybe Int)

The number of hours in the schedule.

scheduledInstanceAvailability_availableInstanceCount :: Lens' ScheduledInstanceAvailability (Maybe Int)

The number of available instances.

scheduledInstanceAvailability_minTermDurationInDays :: Lens' ScheduledInstanceAvailability (Maybe Int)

The minimum term. The only possible value is 365 days.

scheduledInstanceAvailability_availabilityZone :: Lens' ScheduledInstanceAvailability (Maybe Text)

The Availability Zone.

scheduledInstanceAvailability_maxTermDurationInDays :: Lens' ScheduledInstanceAvailability (Maybe Int)

The maximum term. The only possible value is 365 days.

scheduledInstanceAvailability_recurrence :: Lens' ScheduledInstanceAvailability (Maybe ScheduledInstanceRecurrence)

The schedule recurrence.

scheduledInstanceAvailability_totalScheduledInstanceHours :: Lens' ScheduledInstanceAvailability (Maybe Int)

The total number of hours for a single instance for the entire term.

scheduledInstanceAvailability_firstSlotStartTime :: Lens' ScheduledInstanceAvailability (Maybe UTCTime)

The time period for the first schedule to start.

scheduledInstanceAvailability_purchaseToken :: Lens' ScheduledInstanceAvailability (Maybe Text)

The purchase token. This token expires in two hours.

scheduledInstanceAvailability_hourlyPrice :: Lens' ScheduledInstanceAvailability (Maybe Text)

The hourly price for a single instance.