Network.AWS.EC2.Types.ScheduledInstancesBlockDeviceMapping

Description

Documentation

data ScheduledInstancesBlockDeviceMapping

Describes a block device mapping for a Scheduled Instance.

See: newScheduledInstancesBlockDeviceMapping smart constructor.

Constructors

ScheduledInstancesBlockDeviceMapping'

Fields

ebs :: Maybe ScheduledInstancesEbs

Parameters used to set up EBS volumes automatically when the instance is launched.

noDevice :: Maybe Text

To omit the device from the block device mapping, specify an empty string.

virtualName :: Maybe Text

The virtual device name (ephemeralN). Instance store volumes are numbered starting from 0. An instance type with two available instance store volumes can specify mappings for ephemeral0 and ephemeral1. The number of available instance store volumes depends on the instance type. After you connect to the instance, you must mount the volume.

Constraints: For M3 instances, you must specify instance store volumes in the block device mapping for the instance. When you launch an M3 instance, we ignore any instance store volumes specified in the block device mapping for the AMI.

deviceName :: Maybe Text

The device name (for example, /dev/sdh or xvdh).

newScheduledInstancesBlockDeviceMapping :: ScheduledInstancesBlockDeviceMapping

Create a value of ScheduledInstancesBlockDeviceMapping with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ebs:ScheduledInstancesBlockDeviceMapping', scheduledInstancesBlockDeviceMapping_ebs - Parameters used to set up EBS volumes automatically when the instance is launched.

$sel:noDevice:ScheduledInstancesBlockDeviceMapping', scheduledInstancesBlockDeviceMapping_noDevice - To omit the device from the block device mapping, specify an empty string.

$sel:virtualName:ScheduledInstancesBlockDeviceMapping', scheduledInstancesBlockDeviceMapping_virtualName - The virtual device name (ephemeralN). Instance store volumes are numbered starting from 0. An instance type with two available instance store volumes can specify mappings for ephemeral0 and ephemeral1. The number of available instance store volumes depends on the instance type. After you connect to the instance, you must mount the volume.

Constraints: For M3 instances, you must specify instance store volumes in the block device mapping for the instance. When you launch an M3 instance, we ignore any instance store volumes specified in the block device mapping for the AMI.

$sel:deviceName:ScheduledInstancesBlockDeviceMapping', scheduledInstancesBlockDeviceMapping_deviceName - The device name (for example, /dev/sdh or xvdh).

scheduledInstancesBlockDeviceMapping_ebs :: Lens' ScheduledInstancesBlockDeviceMapping (Maybe ScheduledInstancesEbs)

Parameters used to set up EBS volumes automatically when the instance is launched.

scheduledInstancesBlockDeviceMapping_noDevice :: Lens' ScheduledInstancesBlockDeviceMapping (Maybe Text)

To omit the device from the block device mapping, specify an empty string.

scheduledInstancesBlockDeviceMapping_virtualName :: Lens' ScheduledInstancesBlockDeviceMapping (Maybe Text)

The virtual device name (ephemeralN). Instance store volumes are numbered starting from 0. An instance type with two available instance store volumes can specify mappings for ephemeral0 and ephemeral1. The number of available instance store volumes depends on the instance type. After you connect to the instance, you must mount the volume.

Constraints: For M3 instances, you must specify instance store volumes in the block device mapping for the instance. When you launch an M3 instance, we ignore any instance store volumes specified in the block device mapping for the AMI.

scheduledInstancesBlockDeviceMapping_deviceName :: Lens' ScheduledInstancesBlockDeviceMapping (Maybe Text)

The device name (for example, /dev/sdh or xvdh).