Network.AWS.EC2.Types.ScheduledInstancesEbs

Description

Documentation

data ScheduledInstancesEbs

Describes an EBS volume for a Scheduled Instance.

See: newScheduledInstancesEbs smart constructor.

Constructors

ScheduledInstancesEbs'

Fields

encrypted :: Maybe Bool

Indicates whether the volume is encrypted. You can attached encrypted volumes only to instances that support them.

deleteOnTermination :: Maybe Bool

Indicates whether the volume is deleted on instance termination.

snapshotId :: Maybe Text

The ID of the snapshot.

volumeType :: Maybe Text

The volume type. gp2 for General Purpose SSD, io1 or io2 for Provisioned IOPS SSD, Throughput Optimized HDD for st1, Cold HDD for sc1, or standard for Magnetic.

Default: gp2

iops :: Maybe Int

The number of I/O operations per second (IOPS) to provision for an io1 or io2 volume, with a maximum ratio of 50 IOPS/GiB for io1, and 500 IOPS/GiB for io2. Range is 100 to 64,000 IOPS for volumes in most Regions. Maximum IOPS of 64,000 is guaranteed only on instances built on the Nitro System. Other instance families guarantee performance up to 32,000 IOPS. For more information, see Amazon EBS volume types in the Amazon EC2 User Guide.

This parameter is valid only for Provisioned IOPS SSD (io1 and io2) volumes.

volumeSize :: Maybe Int

The size of the volume, in GiB.

Default: If you're creating the volume from a snapshot and don't specify a volume size, the default is the snapshot size.

newScheduledInstancesEbs :: ScheduledInstancesEbs

Create a value of ScheduledInstancesEbs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:encrypted:ScheduledInstancesEbs', scheduledInstancesEbs_encrypted - Indicates whether the volume is encrypted. You can attached encrypted volumes only to instances that support them.

$sel:deleteOnTermination:ScheduledInstancesEbs', scheduledInstancesEbs_deleteOnTermination - Indicates whether the volume is deleted on instance termination.

$sel:snapshotId:ScheduledInstancesEbs', scheduledInstancesEbs_snapshotId - The ID of the snapshot.

$sel:volumeType:ScheduledInstancesEbs', scheduledInstancesEbs_volumeType - The volume type. gp2 for General Purpose SSD, io1 or io2 for Provisioned IOPS SSD, Throughput Optimized HDD for st1, Cold HDD for sc1, or standard for Magnetic.

Default: gp2

$sel:iops:ScheduledInstancesEbs', scheduledInstancesEbs_iops - The number of I/O operations per second (IOPS) to provision for an io1 or io2 volume, with a maximum ratio of 50 IOPS/GiB for io1, and 500 IOPS/GiB for io2. Range is 100 to 64,000 IOPS for volumes in most Regions. Maximum IOPS of 64,000 is guaranteed only on instances built on the Nitro System. Other instance families guarantee performance up to 32,000 IOPS. For more information, see Amazon EBS volume types in the Amazon EC2 User Guide.

This parameter is valid only for Provisioned IOPS SSD (io1 and io2) volumes.

$sel:volumeSize:ScheduledInstancesEbs', scheduledInstancesEbs_volumeSize - The size of the volume, in GiB.

Default: If you're creating the volume from a snapshot and don't specify a volume size, the default is the snapshot size.

scheduledInstancesEbs_encrypted :: Lens' ScheduledInstancesEbs (Maybe Bool)

Indicates whether the volume is encrypted. You can attached encrypted volumes only to instances that support them.

scheduledInstancesEbs_deleteOnTermination :: Lens' ScheduledInstancesEbs (Maybe Bool)

Indicates whether the volume is deleted on instance termination.

scheduledInstancesEbs_snapshotId :: Lens' ScheduledInstancesEbs (Maybe Text)

The ID of the snapshot.

scheduledInstancesEbs_volumeType :: Lens' ScheduledInstancesEbs (Maybe Text)

The volume type. gp2 for General Purpose SSD, io1 or io2 for Provisioned IOPS SSD, Throughput Optimized HDD for st1, Cold HDD for sc1, or standard for Magnetic.

Default: gp2

scheduledInstancesEbs_iops :: Lens' ScheduledInstancesEbs (Maybe Int)

The number of I/O operations per second (IOPS) to provision for an io1 or io2 volume, with a maximum ratio of 50 IOPS/GiB for io1, and 500 IOPS/GiB for io2. Range is 100 to 64,000 IOPS for volumes in most Regions. Maximum IOPS of 64,000 is guaranteed only on instances built on the Nitro System. Other instance families guarantee performance up to 32,000 IOPS. For more information, see Amazon EBS volume types in the Amazon EC2 User Guide.

This parameter is valid only for Provisioned IOPS SSD (io1 and io2) volumes.

scheduledInstancesEbs_volumeSize :: Lens' ScheduledInstancesEbs (Maybe Int)

The size of the volume, in GiB.

Default: If you're creating the volume from a snapshot and don't specify a volume size, the default is the snapshot size.