Network.AWS.EC2.Types.ScheduledInstancesLaunchSpecification

Description

Documentation

data ScheduledInstancesLaunchSpecification

Describes the launch specification for a Scheduled Instance.

If you are launching the Scheduled Instance in EC2-VPC, you must specify the ID of the subnet. You can specify the subnet using either SubnetId or NetworkInterface.

See: newScheduledInstancesLaunchSpecification smart constructor.

Constructors

ScheduledInstancesLaunchSpecification'

Fields

securityGroupIds :: Maybe [Text]

The IDs of the security groups.

ebsOptimized :: Maybe Bool

Indicates whether the instances are optimized for EBS I/O. This optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal EBS I/O performance. This optimization isn't available with all instance types. Additional usage charges apply when using an EBS-optimized instance.

Default: false

placement :: Maybe ScheduledInstancesPlacement

The placement information.

userData :: Maybe Text

The base64-encoded MIME user data.

instanceType :: Maybe Text

The instance type.

ramdiskId :: Maybe Text

The ID of the RAM disk.

iamInstanceProfile :: Maybe ScheduledInstancesIamInstanceProfile

The IAM instance profile.

monitoring :: Maybe ScheduledInstancesMonitoring

Enable or disable monitoring for the instances.

blockDeviceMappings :: Maybe [ScheduledInstancesBlockDeviceMapping]

The block device mapping entries.

subnetId :: Maybe Text

The ID of the subnet in which to launch the instances.

kernelId :: Maybe Text

The ID of the kernel.

networkInterfaces :: Maybe [ScheduledInstancesNetworkInterface]

The network interfaces.

keyName :: Maybe Text

The name of the key pair.

imageId :: Text

The ID of the Amazon Machine Image (AMI).

newScheduledInstancesLaunchSpecification

Arguments

:: Text
$sel:imageId:ScheduledInstancesLaunchSpecification'

-> ScheduledInstancesLaunchSpecification

Create a value of ScheduledInstancesLaunchSpecification with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:securityGroupIds:ScheduledInstancesLaunchSpecification', scheduledInstancesLaunchSpecification_securityGroupIds - The IDs of the security groups.

$sel:ebsOptimized:ScheduledInstancesLaunchSpecification', scheduledInstancesLaunchSpecification_ebsOptimized - Indicates whether the instances are optimized for EBS I/O. This optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal EBS I/O performance. This optimization isn't available with all instance types. Additional usage charges apply when using an EBS-optimized instance.

Default: false

$sel:placement:ScheduledInstancesLaunchSpecification', scheduledInstancesLaunchSpecification_placement - The placement information.

$sel:userData:ScheduledInstancesLaunchSpecification', scheduledInstancesLaunchSpecification_userData - The base64-encoded MIME user data.

$sel:instanceType:ScheduledInstancesLaunchSpecification', scheduledInstancesLaunchSpecification_instanceType - The instance type.

$sel:ramdiskId:ScheduledInstancesLaunchSpecification', scheduledInstancesLaunchSpecification_ramdiskId - The ID of the RAM disk.

$sel:iamInstanceProfile:ScheduledInstancesLaunchSpecification', scheduledInstancesLaunchSpecification_iamInstanceProfile - The IAM instance profile.

$sel:monitoring:ScheduledInstancesLaunchSpecification', scheduledInstancesLaunchSpecification_monitoring - Enable or disable monitoring for the instances.

$sel:blockDeviceMappings:ScheduledInstancesLaunchSpecification', scheduledInstancesLaunchSpecification_blockDeviceMappings - The block device mapping entries.

$sel:subnetId:ScheduledInstancesLaunchSpecification', scheduledInstancesLaunchSpecification_subnetId - The ID of the subnet in which to launch the instances.

$sel:kernelId:ScheduledInstancesLaunchSpecification', scheduledInstancesLaunchSpecification_kernelId - The ID of the kernel.

$sel:networkInterfaces:ScheduledInstancesLaunchSpecification', scheduledInstancesLaunchSpecification_networkInterfaces - The network interfaces.

$sel:keyName:ScheduledInstancesLaunchSpecification', scheduledInstancesLaunchSpecification_keyName - The name of the key pair.

$sel:imageId:ScheduledInstancesLaunchSpecification', scheduledInstancesLaunchSpecification_imageId - The ID of the Amazon Machine Image (AMI).

scheduledInstancesLaunchSpecification_securityGroupIds :: Lens' ScheduledInstancesLaunchSpecification (Maybe [Text])

The IDs of the security groups.

scheduledInstancesLaunchSpecification_ebsOptimized :: Lens' ScheduledInstancesLaunchSpecification (Maybe Bool)

Indicates whether the instances are optimized for EBS I/O. This optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal EBS I/O performance. This optimization isn't available with all instance types. Additional usage charges apply when using an EBS-optimized instance.

Default: false

scheduledInstancesLaunchSpecification_placement :: Lens' ScheduledInstancesLaunchSpecification (Maybe ScheduledInstancesPlacement)

The placement information.

scheduledInstancesLaunchSpecification_userData :: Lens' ScheduledInstancesLaunchSpecification (Maybe Text)

The base64-encoded MIME user data.

scheduledInstancesLaunchSpecification_instanceType :: Lens' ScheduledInstancesLaunchSpecification (Maybe Text)

The instance type.

scheduledInstancesLaunchSpecification_ramdiskId :: Lens' ScheduledInstancesLaunchSpecification (Maybe Text)

The ID of the RAM disk.

scheduledInstancesLaunchSpecification_iamInstanceProfile :: Lens' ScheduledInstancesLaunchSpecification (Maybe ScheduledInstancesIamInstanceProfile)

The IAM instance profile.

scheduledInstancesLaunchSpecification_monitoring :: Lens' ScheduledInstancesLaunchSpecification (Maybe ScheduledInstancesMonitoring)

Enable or disable monitoring for the instances.

scheduledInstancesLaunchSpecification_blockDeviceMappings :: Lens' ScheduledInstancesLaunchSpecification (Maybe [ScheduledInstancesBlockDeviceMapping])

The block device mapping entries.

scheduledInstancesLaunchSpecification_subnetId :: Lens' ScheduledInstancesLaunchSpecification (Maybe Text)

The ID of the subnet in which to launch the instances.

scheduledInstancesLaunchSpecification_kernelId :: Lens' ScheduledInstancesLaunchSpecification (Maybe Text)

The ID of the kernel.

scheduledInstancesLaunchSpecification_networkInterfaces :: Lens' ScheduledInstancesLaunchSpecification (Maybe [ScheduledInstancesNetworkInterface])

The network interfaces.

scheduledInstancesLaunchSpecification_keyName :: Lens' ScheduledInstancesLaunchSpecification (Maybe Text)

The name of the key pair.

scheduledInstancesLaunchSpecification_imageId :: Lens' ScheduledInstancesLaunchSpecification Text

The ID of the Amazon Machine Image (AMI).