Network.AWS.EC2.Types.SecurityGroupReference

Description

Documentation

data SecurityGroupReference

Describes a VPC with a security group that references your security group.

See: newSecurityGroupReference smart constructor.

Constructors

SecurityGroupReference'

Fields

vpcPeeringConnectionId :: Maybe Text

The ID of the VPC peering connection.

groupId :: Maybe Text

The ID of your security group.

referencingVpcId :: Maybe Text

The ID of the VPC with the referencing security group.

newSecurityGroupReference :: SecurityGroupReference

Create a value of SecurityGroupReference with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpcPeeringConnectionId:SecurityGroupReference', securityGroupReference_vpcPeeringConnectionId - The ID of the VPC peering connection.

$sel:groupId:SecurityGroupReference', securityGroupReference_groupId - The ID of your security group.

$sel:referencingVpcId:SecurityGroupReference', securityGroupReference_referencingVpcId - The ID of the VPC with the referencing security group.

securityGroupReference_vpcPeeringConnectionId :: Lens' SecurityGroupReference (Maybe Text)

The ID of the VPC peering connection.

securityGroupReference_groupId :: Lens' SecurityGroupReference (Maybe Text)

The ID of your security group.

securityGroupReference_referencingVpcId :: Lens' SecurityGroupReference (Maybe Text)

The ID of the VPC with the referencing security group.