Network.AWS.EC2.Types.SecurityGroupRuleRequest

Description

Documentation

data SecurityGroupRuleRequest

Describes a security group rule.

You must specify exactly one of the following parameters, based on the rule type:

CidrIpv4

CidrIpv6

PrefixListId

ReferencedGroupId

When you modify a rule, you cannot change the rule type. For example, if the rule uses an IPv4 address range, you must use CidrIpv4 to specify a new IPv4 address range.

See: newSecurityGroupRuleRequest smart constructor.

Constructors

SecurityGroupRuleRequest'

Fields

cidrIpv4 :: Maybe Text

The IPv4 CIDR range. To specify a single IPv4 address, use the /32 prefix length.

fromPort :: Maybe Int

The start of port range for the TCP and UDP protocols, or an ICMP/ICMPv6 type. A value of -1 indicates all ICMP/ICMPv6 types. If you specify all ICMP/ICMPv6 types, you must specify all codes.

referencedGroupId :: Maybe Text

The ID of the security group that is referenced in the security group rule.

prefixListId :: Maybe Text

The ID of the prefix list.

ipProtocol :: Maybe Text

The IP protocol name (tcp, udp, icmp, icmpv6) or number (see Protocol Numbers).

Use -1 to specify all protocols.

toPort :: Maybe Int

The end of port range for the TCP and UDP protocols, or an ICMP/ICMPv6 code. A value of -1 indicates all ICMP/ICMPv6 codes. If you specify all ICMP/ICMPv6 types, you must specify all codes.

cidrIpv6 :: Maybe Text

The IPv6 CIDR range. To specify a single IPv6 address, use the /128 prefix length.

description :: Maybe Text

The description of the security group rule.

newSecurityGroupRuleRequest :: SecurityGroupRuleRequest

Create a value of SecurityGroupRuleRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cidrIpv4:SecurityGroupRuleRequest', securityGroupRuleRequest_cidrIpv4 - The IPv4 CIDR range. To specify a single IPv4 address, use the /32 prefix length.

$sel:fromPort:SecurityGroupRuleRequest', securityGroupRuleRequest_fromPort - The start of port range for the TCP and UDP protocols, or an ICMP/ICMPv6 type. A value of -1 indicates all ICMP/ICMPv6 types. If you specify all ICMP/ICMPv6 types, you must specify all codes.

$sel:referencedGroupId:SecurityGroupRuleRequest', securityGroupRuleRequest_referencedGroupId - The ID of the security group that is referenced in the security group rule.

$sel:prefixListId:SecurityGroupRuleRequest', securityGroupRuleRequest_prefixListId - The ID of the prefix list.

$sel:ipProtocol:SecurityGroupRuleRequest', securityGroupRuleRequest_ipProtocol - The IP protocol name (tcp, udp, icmp, icmpv6) or number (see Protocol Numbers).

Use -1 to specify all protocols.

$sel:toPort:SecurityGroupRuleRequest', securityGroupRuleRequest_toPort - The end of port range for the TCP and UDP protocols, or an ICMP/ICMPv6 code. A value of -1 indicates all ICMP/ICMPv6 codes. If you specify all ICMP/ICMPv6 types, you must specify all codes.

$sel:cidrIpv6:SecurityGroupRuleRequest', securityGroupRuleRequest_cidrIpv6 - The IPv6 CIDR range. To specify a single IPv6 address, use the /128 prefix length.

$sel:description:SecurityGroupRuleRequest', securityGroupRuleRequest_description - The description of the security group rule.

securityGroupRuleRequest_cidrIpv4 :: Lens' SecurityGroupRuleRequest (Maybe Text)

The IPv4 CIDR range. To specify a single IPv4 address, use the /32 prefix length.

securityGroupRuleRequest_fromPort :: Lens' SecurityGroupRuleRequest (Maybe Int)

The start of port range for the TCP and UDP protocols, or an ICMP/ICMPv6 type. A value of -1 indicates all ICMP/ICMPv6 types. If you specify all ICMP/ICMPv6 types, you must specify all codes.

securityGroupRuleRequest_referencedGroupId :: Lens' SecurityGroupRuleRequest (Maybe Text)

The ID of the security group that is referenced in the security group rule.

securityGroupRuleRequest_prefixListId :: Lens' SecurityGroupRuleRequest (Maybe Text)

The ID of the prefix list.

securityGroupRuleRequest_ipProtocol :: Lens' SecurityGroupRuleRequest (Maybe Text)

The IP protocol name (tcp, udp, icmp, icmpv6) or number (see Protocol Numbers).

Use -1 to specify all protocols.

securityGroupRuleRequest_toPort :: Lens' SecurityGroupRuleRequest (Maybe Int)

The end of port range for the TCP and UDP protocols, or an ICMP/ICMPv6 code. A value of -1 indicates all ICMP/ICMPv6 codes. If you specify all ICMP/ICMPv6 types, you must specify all codes.

securityGroupRuleRequest_cidrIpv6 :: Lens' SecurityGroupRuleRequest (Maybe Text)

The IPv6 CIDR range. To specify a single IPv6 address, use the /128 prefix length.

securityGroupRuleRequest_description :: Lens' SecurityGroupRuleRequest (Maybe Text)

The description of the security group rule.