Network.AWS.EC2.Types.ServiceConfiguration

Description

Documentation

data ServiceConfiguration

Describes a service configuration for a VPC endpoint service.

See: newServiceConfiguration smart constructor.

Constructors

ServiceConfiguration'

Fields

managesVpcEndpoints :: Maybe Bool

Indicates whether the service manages its VPC endpoints. Management of the service VPC endpoints using the VPC endpoint API is restricted.

availabilityZones :: Maybe [Text]

The Availability Zones in which the service is available.

gatewayLoadBalancerArns :: Maybe [Text]

The Amazon Resource Names (ARNs) of the Gateway Load Balancers for the service.

privateDnsNameConfiguration :: Maybe PrivateDnsNameConfiguration

Information about the endpoint service private DNS name configuration.

baseEndpointDnsNames :: Maybe [Text]

The DNS names for the service.

serviceId :: Maybe Text

The ID of the service.

serviceName :: Maybe Text

The name of the service.

tags :: Maybe [Tag]

Any tags assigned to the service.

privateDnsName :: Maybe Text

The private DNS name for the service.

acceptanceRequired :: Maybe Bool

Indicates whether requests from other AWS accounts to create an endpoint to the service must first be accepted.

serviceType :: Maybe [ServiceTypeDetail]

The type of service.

serviceState :: Maybe ServiceState

The service state.

networkLoadBalancerArns :: Maybe [Text]

The Amazon Resource Names (ARNs) of the Network Load Balancers for the service.

newServiceConfiguration :: ServiceConfiguration

Create a value of ServiceConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:managesVpcEndpoints:ServiceConfiguration', serviceConfiguration_managesVpcEndpoints - Indicates whether the service manages its VPC endpoints. Management of the service VPC endpoints using the VPC endpoint API is restricted.

$sel:availabilityZones:ServiceConfiguration', serviceConfiguration_availabilityZones - The Availability Zones in which the service is available.

$sel:gatewayLoadBalancerArns:ServiceConfiguration', serviceConfiguration_gatewayLoadBalancerArns - The Amazon Resource Names (ARNs) of the Gateway Load Balancers for the service.

$sel:privateDnsNameConfiguration:ServiceConfiguration', serviceConfiguration_privateDnsNameConfiguration - Information about the endpoint service private DNS name configuration.

$sel:baseEndpointDnsNames:ServiceConfiguration', serviceConfiguration_baseEndpointDnsNames - The DNS names for the service.

$sel:serviceId:ServiceConfiguration', serviceConfiguration_serviceId - The ID of the service.

$sel:serviceName:ServiceConfiguration', serviceConfiguration_serviceName - The name of the service.

$sel:tags:ServiceConfiguration', serviceConfiguration_tags - Any tags assigned to the service.

$sel:privateDnsName:ServiceConfiguration', serviceConfiguration_privateDnsName - The private DNS name for the service.

$sel:acceptanceRequired:ServiceConfiguration', serviceConfiguration_acceptanceRequired - Indicates whether requests from other AWS accounts to create an endpoint to the service must first be accepted.

$sel:serviceType:ServiceConfiguration', serviceConfiguration_serviceType - The type of service.

$sel:serviceState:ServiceConfiguration', serviceConfiguration_serviceState - The service state.

$sel:networkLoadBalancerArns:ServiceConfiguration', serviceConfiguration_networkLoadBalancerArns - The Amazon Resource Names (ARNs) of the Network Load Balancers for the service.

serviceConfiguration_managesVpcEndpoints :: Lens' ServiceConfiguration (Maybe Bool)

Indicates whether the service manages its VPC endpoints. Management of the service VPC endpoints using the VPC endpoint API is restricted.

serviceConfiguration_availabilityZones :: Lens' ServiceConfiguration (Maybe [Text])

The Availability Zones in which the service is available.

serviceConfiguration_gatewayLoadBalancerArns :: Lens' ServiceConfiguration (Maybe [Text])

The Amazon Resource Names (ARNs) of the Gateway Load Balancers for the service.

serviceConfiguration_privateDnsNameConfiguration :: Lens' ServiceConfiguration (Maybe PrivateDnsNameConfiguration)

Information about the endpoint service private DNS name configuration.

serviceConfiguration_baseEndpointDnsNames :: Lens' ServiceConfiguration (Maybe [Text])

The DNS names for the service.

serviceConfiguration_serviceId :: Lens' ServiceConfiguration (Maybe Text)

The ID of the service.

serviceConfiguration_serviceName :: Lens' ServiceConfiguration (Maybe Text)

The name of the service.

serviceConfiguration_tags :: Lens' ServiceConfiguration (Maybe [Tag])

Any tags assigned to the service.

serviceConfiguration_privateDnsName :: Lens' ServiceConfiguration (Maybe Text)

The private DNS name for the service.

serviceConfiguration_acceptanceRequired :: Lens' ServiceConfiguration (Maybe Bool)

Indicates whether requests from other AWS accounts to create an endpoint to the service must first be accepted.

serviceConfiguration_serviceType :: Lens' ServiceConfiguration (Maybe [ServiceTypeDetail])

The type of service.

serviceConfiguration_serviceState :: Lens' ServiceConfiguration (Maybe ServiceState)

The service state.

serviceConfiguration_networkLoadBalancerArns :: Lens' ServiceConfiguration (Maybe [Text])

The Amazon Resource Names (ARNs) of the Network Load Balancers for the service.