Network.AWS.EC2.Types.ServiceDetail

Description

Documentation

data ServiceDetail

Describes a VPC endpoint service.

See: newServiceDetail smart constructor.

Constructors

ServiceDetail'

Fields

managesVpcEndpoints :: Maybe Bool

Indicates whether the service manages its VPC endpoints. Management of the service VPC endpoints using the VPC endpoint API is restricted.

availabilityZones :: Maybe [Text]

The Availability Zones in which the service is available.

baseEndpointDnsNames :: Maybe [Text]

The DNS names for the service.

serviceId :: Maybe Text

The ID of the endpoint service.

privateDnsNames :: Maybe [PrivateDnsDetails]

The private DNS names assigned to the VPC endpoint service.

vpcEndpointPolicySupported :: Maybe Bool

Indicates whether the service supports endpoint policies.

privateDnsNameVerificationState :: Maybe DnsNameState

The verification state of the VPC endpoint service.

Consumers of the endpoint service cannot use the private name when the state is not verified.

serviceName :: Maybe Text

The Amazon Resource Name (ARN) of the service.

tags :: Maybe [Tag]

Any tags assigned to the service.

owner :: Maybe Text

The AWS account ID of the service owner.

privateDnsName :: Maybe Text

The private DNS name for the service.

acceptanceRequired :: Maybe Bool

Indicates whether VPC endpoint connection requests to the service must be accepted by the service owner.

serviceType :: Maybe [ServiceTypeDetail]

The type of service.

newServiceDetail :: ServiceDetail

Create a value of ServiceDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:managesVpcEndpoints:ServiceDetail', serviceDetail_managesVpcEndpoints - Indicates whether the service manages its VPC endpoints. Management of the service VPC endpoints using the VPC endpoint API is restricted.

$sel:availabilityZones:ServiceDetail', serviceDetail_availabilityZones - The Availability Zones in which the service is available.

$sel:baseEndpointDnsNames:ServiceDetail', serviceDetail_baseEndpointDnsNames - The DNS names for the service.

$sel:serviceId:ServiceDetail', serviceDetail_serviceId - The ID of the endpoint service.

$sel:privateDnsNames:ServiceDetail', serviceDetail_privateDnsNames - The private DNS names assigned to the VPC endpoint service.

$sel:vpcEndpointPolicySupported:ServiceDetail', serviceDetail_vpcEndpointPolicySupported - Indicates whether the service supports endpoint policies.

$sel:privateDnsNameVerificationState:ServiceDetail', serviceDetail_privateDnsNameVerificationState - The verification state of the VPC endpoint service.

Consumers of the endpoint service cannot use the private name when the state is not verified.

$sel:serviceName:ServiceDetail', serviceDetail_serviceName - The Amazon Resource Name (ARN) of the service.

$sel:tags:ServiceDetail', serviceDetail_tags - Any tags assigned to the service.

$sel:owner:ServiceDetail', serviceDetail_owner - The AWS account ID of the service owner.

$sel:privateDnsName:ServiceDetail', serviceDetail_privateDnsName - The private DNS name for the service.

$sel:acceptanceRequired:ServiceDetail', serviceDetail_acceptanceRequired - Indicates whether VPC endpoint connection requests to the service must be accepted by the service owner.

$sel:serviceType:ServiceDetail', serviceDetail_serviceType - The type of service.

serviceDetail_managesVpcEndpoints :: Lens' ServiceDetail (Maybe Bool)

Indicates whether the service manages its VPC endpoints. Management of the service VPC endpoints using the VPC endpoint API is restricted.

serviceDetail_availabilityZones :: Lens' ServiceDetail (Maybe [Text])

The Availability Zones in which the service is available.

serviceDetail_baseEndpointDnsNames :: Lens' ServiceDetail (Maybe [Text])

The DNS names for the service.

serviceDetail_serviceId :: Lens' ServiceDetail (Maybe Text)

The ID of the endpoint service.

serviceDetail_privateDnsNames :: Lens' ServiceDetail (Maybe [PrivateDnsDetails])

The private DNS names assigned to the VPC endpoint service.

serviceDetail_vpcEndpointPolicySupported :: Lens' ServiceDetail (Maybe Bool)

Indicates whether the service supports endpoint policies.

serviceDetail_privateDnsNameVerificationState :: Lens' ServiceDetail (Maybe DnsNameState)

The verification state of the VPC endpoint service.

Consumers of the endpoint service cannot use the private name when the state is not verified.

serviceDetail_serviceName :: Lens' ServiceDetail (Maybe Text)

The Amazon Resource Name (ARN) of the service.

serviceDetail_tags :: Lens' ServiceDetail (Maybe [Tag])

Any tags assigned to the service.

serviceDetail_owner :: Lens' ServiceDetail (Maybe Text)

The AWS account ID of the service owner.

serviceDetail_privateDnsName :: Lens' ServiceDetail (Maybe Text)

The private DNS name for the service.

serviceDetail_acceptanceRequired :: Lens' ServiceDetail (Maybe Bool)

Indicates whether VPC endpoint connection requests to the service must be accepted by the service owner.

serviceDetail_serviceType :: Lens' ServiceDetail (Maybe [ServiceTypeDetail])

The type of service.