Network.AWS.EC2.Types.Snapshot

Description

Documentation

data Snapshot

Describes a snapshot.

See: newSnapshot smart constructor.

Constructors

Snapshot'

Fields

ownerAlias :: Maybe Text

The Amazon Web Services owner alias, from an Amazon-maintained list (amazon). This is not the user-configured Amazon Web Services account alias set using the IAM console.

stateMessage :: Maybe Text

Encrypted Amazon EBS snapshots are copied asynchronously. If a snapshot copy operation fails (for example, if the proper Key Management Service (KMS) permissions are not obtained) this field displays error state details to help you diagnose why the error occurred. This parameter is only returned by DescribeSnapshots.

outpostArn :: Maybe Text

The ARN of the Outpost on which the snapshot is stored. For more information, see Amazon EBS local snapshots on Outposts in the Amazon Elastic Compute Cloud User Guide.

dataEncryptionKeyId :: Maybe Text

The data encryption key identifier for the snapshot. This value is a unique identifier that corresponds to the data encryption key that was used to encrypt the original volume or snapshot copy. Because data encryption keys are inherited by volumes created from snapshots, and vice versa, if snapshots share the same data encryption key identifier, then they belong to the same volume/snapshot lineage. This parameter is only returned by DescribeSnapshots.

kmsKeyId :: Maybe Text

The Amazon Resource Name (ARN) of the Key Management Service (KMS) KMS key that was used to protect the volume encryption key for the parent volume.

tags :: Maybe [Tag]

Any tags assigned to the snapshot.

snapshotId :: Text

The ID of the snapshot. Each snapshot receives a unique identifier when it is created.

ownerId :: Text

The ID of the Amazon Web Services account that owns the EBS snapshot.

volumeId :: Text

The ID of the volume that was used to create the snapshot. Snapshots created by the CopySnapshot action have an arbitrary volume ID that should not be used for any purpose.

volumeSize :: Int

The size of the volume, in GiB.

description :: Text

The description for the snapshot.

startTime :: ISO8601

The time stamp when the snapshot was initiated.

progress :: Text

The progress of the snapshot, as a percentage.

state :: SnapshotState

The snapshot state.

encrypted :: Bool

Indicates whether the snapshot is encrypted.

newSnapshot

Arguments

:: Text
$sel:snapshotId:Snapshot'

-> Text
$sel:ownerId:Snapshot'

-> Text
$sel:volumeId:Snapshot'

-> Int
$sel:volumeSize:Snapshot'

-> Text
$sel:description:Snapshot'

-> UTCTime
$sel:startTime:Snapshot'

-> Text
$sel:progress:Snapshot'

-> SnapshotState
$sel:state:Snapshot'

-> Bool
$sel:encrypted:Snapshot'

-> Snapshot

Create a value of Snapshot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ownerAlias:Snapshot', snapshot_ownerAlias - The Amazon Web Services owner alias, from an Amazon-maintained list (amazon). This is not the user-configured Amazon Web Services account alias set using the IAM console.

$sel:stateMessage:Snapshot', snapshot_stateMessage - Encrypted Amazon EBS snapshots are copied asynchronously. If a snapshot copy operation fails (for example, if the proper Key Management Service (KMS) permissions are not obtained) this field displays error state details to help you diagnose why the error occurred. This parameter is only returned by DescribeSnapshots.

$sel:outpostArn:Snapshot', snapshot_outpostArn - The ARN of the Outpost on which the snapshot is stored. For more information, see Amazon EBS local snapshots on Outposts in the Amazon Elastic Compute Cloud User Guide.

$sel:dataEncryptionKeyId:Snapshot', snapshot_dataEncryptionKeyId - The data encryption key identifier for the snapshot. This value is a unique identifier that corresponds to the data encryption key that was used to encrypt the original volume or snapshot copy. Because data encryption keys are inherited by volumes created from snapshots, and vice versa, if snapshots share the same data encryption key identifier, then they belong to the same volume/snapshot lineage. This parameter is only returned by DescribeSnapshots.

$sel:kmsKeyId:Snapshot', snapshot_kmsKeyId - The Amazon Resource Name (ARN) of the Key Management Service (KMS) KMS key that was used to protect the volume encryption key for the parent volume.

$sel:tags:Snapshot', snapshot_tags - Any tags assigned to the snapshot.

$sel:snapshotId:Snapshot', snapshot_snapshotId - The ID of the snapshot. Each snapshot receives a unique identifier when it is created.

$sel:ownerId:Snapshot', snapshot_ownerId - The ID of the Amazon Web Services account that owns the EBS snapshot.

$sel:volumeId:Snapshot', snapshot_volumeId - The ID of the volume that was used to create the snapshot. Snapshots created by the CopySnapshot action have an arbitrary volume ID that should not be used for any purpose.

$sel:volumeSize:Snapshot', snapshot_volumeSize - The size of the volume, in GiB.

$sel:description:Snapshot', snapshot_description - The description for the snapshot.

$sel:startTime:Snapshot', snapshot_startTime - The time stamp when the snapshot was initiated.

$sel:progress:Snapshot', snapshot_progress - The progress of the snapshot, as a percentage.

$sel:state:Snapshot', snapshot_state - The snapshot state.

$sel:encrypted:Snapshot', snapshot_encrypted - Indicates whether the snapshot is encrypted.

snapshot_ownerAlias :: Lens' Snapshot (Maybe Text)

The Amazon Web Services owner alias, from an Amazon-maintained list (amazon). This is not the user-configured Amazon Web Services account alias set using the IAM console.

snapshot_stateMessage :: Lens' Snapshot (Maybe Text)

Encrypted Amazon EBS snapshots are copied asynchronously. If a snapshot copy operation fails (for example, if the proper Key Management Service (KMS) permissions are not obtained) this field displays error state details to help you diagnose why the error occurred. This parameter is only returned by DescribeSnapshots.

snapshot_outpostArn :: Lens' Snapshot (Maybe Text)

The ARN of the Outpost on which the snapshot is stored. For more information, see Amazon EBS local snapshots on Outposts in the Amazon Elastic Compute Cloud User Guide.

snapshot_dataEncryptionKeyId :: Lens' Snapshot (Maybe Text)

The data encryption key identifier for the snapshot. This value is a unique identifier that corresponds to the data encryption key that was used to encrypt the original volume or snapshot copy. Because data encryption keys are inherited by volumes created from snapshots, and vice versa, if snapshots share the same data encryption key identifier, then they belong to the same volume/snapshot lineage. This parameter is only returned by DescribeSnapshots.

snapshot_kmsKeyId :: Lens' Snapshot (Maybe Text)

The Amazon Resource Name (ARN) of the Key Management Service (KMS) KMS key that was used to protect the volume encryption key for the parent volume.

snapshot_tags :: Lens' Snapshot (Maybe [Tag])

Any tags assigned to the snapshot.

snapshot_snapshotId :: Lens' Snapshot Text

The ID of the snapshot. Each snapshot receives a unique identifier when it is created.

snapshot_ownerId :: Lens' Snapshot Text

The ID of the Amazon Web Services account that owns the EBS snapshot.

snapshot_volumeId :: Lens' Snapshot Text

The ID of the volume that was used to create the snapshot. Snapshots created by the CopySnapshot action have an arbitrary volume ID that should not be used for any purpose.

snapshot_volumeSize :: Lens' Snapshot Int

The size of the volume, in GiB.

snapshot_description :: Lens' Snapshot Text

The description for the snapshot.

snapshot_startTime :: Lens' Snapshot UTCTime

The time stamp when the snapshot was initiated.

snapshot_progress :: Lens' Snapshot Text

The progress of the snapshot, as a percentage.

snapshot_state :: Lens' Snapshot SnapshotState

The snapshot state.

snapshot_encrypted :: Lens' Snapshot Bool

Indicates whether the snapshot is encrypted.