Network.AWS.EC2.Types.SnapshotDetail

Description

Documentation

data SnapshotDetail

Describes the snapshot created from the imported disk.

See: newSnapshotDetail smart constructor.

Constructors

SnapshotDetail'

Fields

diskImageSize :: Maybe Double

The size of the disk in the snapshot, in GiB.

statusMessage :: Maybe Text

A detailed status message for the snapshot creation.

status :: Maybe Text

A brief status of the snapshot creation.

format :: Maybe Text

The format of the disk image from which the snapshot is created.

deviceName :: Maybe Text

The block device mapping for the snapshot.

userBucket :: Maybe UserBucketDetails

The Amazon S3 bucket for the disk image.

snapshotId :: Maybe Text

The snapshot ID of the disk being imported.

description :: Maybe Text

A description for the snapshot.

url :: Maybe Text

The URL used to access the disk image.

progress :: Maybe Text

The percentage of progress for the task.

newSnapshotDetail :: SnapshotDetail

Create a value of SnapshotDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:diskImageSize:SnapshotDetail', snapshotDetail_diskImageSize - The size of the disk in the snapshot, in GiB.

$sel:statusMessage:SnapshotDetail', snapshotDetail_statusMessage - A detailed status message for the snapshot creation.

$sel:status:SnapshotDetail', snapshotDetail_status - A brief status of the snapshot creation.

$sel:format:SnapshotDetail', snapshotDetail_format - The format of the disk image from which the snapshot is created.

$sel:deviceName:SnapshotDetail', snapshotDetail_deviceName - The block device mapping for the snapshot.

$sel:userBucket:SnapshotDetail', snapshotDetail_userBucket - The Amazon S3 bucket for the disk image.

$sel:snapshotId:SnapshotDetail', snapshotDetail_snapshotId - The snapshot ID of the disk being imported.

$sel:description:SnapshotDetail', snapshotDetail_description - A description for the snapshot.

$sel:url:SnapshotDetail', snapshotDetail_url - The URL used to access the disk image.

$sel:progress:SnapshotDetail', snapshotDetail_progress - The percentage of progress for the task.

snapshotDetail_diskImageSize :: Lens' SnapshotDetail (Maybe Double)

The size of the disk in the snapshot, in GiB.

snapshotDetail_statusMessage :: Lens' SnapshotDetail (Maybe Text)

A detailed status message for the snapshot creation.

snapshotDetail_status :: Lens' SnapshotDetail (Maybe Text)

A brief status of the snapshot creation.

snapshotDetail_format :: Lens' SnapshotDetail (Maybe Text)

The format of the disk image from which the snapshot is created.

snapshotDetail_deviceName :: Lens' SnapshotDetail (Maybe Text)

The block device mapping for the snapshot.

snapshotDetail_userBucket :: Lens' SnapshotDetail (Maybe UserBucketDetails)

The Amazon S3 bucket for the disk image.

snapshotDetail_snapshotId :: Lens' SnapshotDetail (Maybe Text)

The snapshot ID of the disk being imported.

snapshotDetail_description :: Lens' SnapshotDetail (Maybe Text)

A description for the snapshot.

snapshotDetail_url :: Lens' SnapshotDetail (Maybe Text)

The URL used to access the disk image.

snapshotDetail_progress :: Lens' SnapshotDetail (Maybe Text)

The percentage of progress for the task.