Network.AWS.EC2.Types.SnapshotInfo

Description

Documentation

data SnapshotInfo

Information about a snapshot.

See: newSnapshotInfo smart constructor.

Constructors

SnapshotInfo'

Fields

ownerId :: Maybe Text

Account id used when creating this snapshot.

encrypted :: Maybe Bool

Indicates whether the snapshot is encrypted.

outpostArn :: Maybe Text

The ARN of the Outpost on which the snapshot is stored. For more information, see Amazon EBS local snapshots on Outposts in the Amazon Elastic Compute Cloud User Guide.

startTime :: Maybe ISO8601

Time this snapshot was started. This is the same for all snapshots initiated by the same request.

volumeId :: Maybe Text

Source volume from which this snapshot was created.

state :: Maybe SnapshotState

Current state of the snapshot.

tags :: Maybe [Tag]

Tags associated with this snapshot.

snapshotId :: Maybe Text

Snapshot id that can be used to describe this snapshot.

description :: Maybe Text

Description specified by the CreateSnapshotRequest that has been applied to all snapshots.

volumeSize :: Maybe Int

Size of the volume from which this snapshot was created.

progress :: Maybe Text

Progress this snapshot has made towards completing.

newSnapshotInfo :: SnapshotInfo

Create a value of SnapshotInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ownerId:SnapshotInfo', snapshotInfo_ownerId - Account id used when creating this snapshot.

$sel:encrypted:SnapshotInfo', snapshotInfo_encrypted - Indicates whether the snapshot is encrypted.

$sel:outpostArn:SnapshotInfo', snapshotInfo_outpostArn - The ARN of the Outpost on which the snapshot is stored. For more information, see Amazon EBS local snapshots on Outposts in the Amazon Elastic Compute Cloud User Guide.

$sel:startTime:SnapshotInfo', snapshotInfo_startTime - Time this snapshot was started. This is the same for all snapshots initiated by the same request.

$sel:volumeId:SnapshotInfo', snapshotInfo_volumeId - Source volume from which this snapshot was created.

$sel:state:SnapshotInfo', snapshotInfo_state - Current state of the snapshot.

$sel:tags:SnapshotInfo', snapshotInfo_tags - Tags associated with this snapshot.

$sel:snapshotId:SnapshotInfo', snapshotInfo_snapshotId - Snapshot id that can be used to describe this snapshot.

$sel:description:SnapshotInfo', snapshotInfo_description - Description specified by the CreateSnapshotRequest that has been applied to all snapshots.

$sel:volumeSize:SnapshotInfo', snapshotInfo_volumeSize - Size of the volume from which this snapshot was created.

$sel:progress:SnapshotInfo', snapshotInfo_progress - Progress this snapshot has made towards completing.

snapshotInfo_ownerId :: Lens' SnapshotInfo (Maybe Text)

Account id used when creating this snapshot.

snapshotInfo_encrypted :: Lens' SnapshotInfo (Maybe Bool)

Indicates whether the snapshot is encrypted.

snapshotInfo_outpostArn :: Lens' SnapshotInfo (Maybe Text)

The ARN of the Outpost on which the snapshot is stored. For more information, see Amazon EBS local snapshots on Outposts in the Amazon Elastic Compute Cloud User Guide.

snapshotInfo_startTime :: Lens' SnapshotInfo (Maybe UTCTime)

Time this snapshot was started. This is the same for all snapshots initiated by the same request.

snapshotInfo_volumeId :: Lens' SnapshotInfo (Maybe Text)

Source volume from which this snapshot was created.

snapshotInfo_state :: Lens' SnapshotInfo (Maybe SnapshotState)

Current state of the snapshot.

snapshotInfo_tags :: Lens' SnapshotInfo (Maybe [Tag])

Tags associated with this snapshot.

snapshotInfo_snapshotId :: Lens' SnapshotInfo (Maybe Text)

Snapshot id that can be used to describe this snapshot.

snapshotInfo_description :: Lens' SnapshotInfo (Maybe Text)

Description specified by the CreateSnapshotRequest that has been applied to all snapshots.

snapshotInfo_volumeSize :: Lens' SnapshotInfo (Maybe Int)

Size of the volume from which this snapshot was created.

snapshotInfo_progress :: Lens' SnapshotInfo (Maybe Text)

Progress this snapshot has made towards completing.