Network.AWS.EC2.Types.SpotCapacityRebalance

Description

Documentation

data SpotCapacityRebalance

The Spot Instance replacement strategy to use when Amazon EC2 emits a signal that your Spot Instance is at an elevated risk of being interrupted. For more information, see Capacity rebalancing in the Amazon EC2 User Guide for Linux Instances.

See: newSpotCapacityRebalance smart constructor.

Constructors

SpotCapacityRebalance'

Fields

replacementStrategy :: Maybe ReplacementStrategy

The replacement strategy to use. Only available for fleets of type maintain. You must specify a value, otherwise you get an error.

To allow Spot Fleet to launch a replacement Spot Instance when an instance rebalance notification is emitted for a Spot Instance in the fleet, specify launch.

When a replacement instance is launched, the instance marked for rebalance is not automatically terminated. You can terminate it, or you can leave it running. You are charged for all instances while they are running.

newSpotCapacityRebalance :: SpotCapacityRebalance

Create a value of SpotCapacityRebalance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:replacementStrategy:SpotCapacityRebalance', spotCapacityRebalance_replacementStrategy - The replacement strategy to use. Only available for fleets of type maintain. You must specify a value, otherwise you get an error.

To allow Spot Fleet to launch a replacement Spot Instance when an instance rebalance notification is emitted for a Spot Instance in the fleet, specify launch.

When a replacement instance is launched, the instance marked for rebalance is not automatically terminated. You can terminate it, or you can leave it running. You are charged for all instances while they are running.

spotCapacityRebalance_replacementStrategy :: Lens' SpotCapacityRebalance (Maybe ReplacementStrategy)

The replacement strategy to use. Only available for fleets of type maintain. You must specify a value, otherwise you get an error.

To allow Spot Fleet to launch a replacement Spot Instance when an instance rebalance notification is emitted for a Spot Instance in the fleet, specify launch.

When a replacement instance is launched, the instance marked for rebalance is not automatically terminated. You can terminate it, or you can leave it running. You are charged for all instances while they are running.