Network.AWS.EC2.Types.SpotDatafeedSubscription

Description

Documentation

data SpotDatafeedSubscription

Describes the data feed for a Spot Instance.

See: newSpotDatafeedSubscription smart constructor.

Constructors

SpotDatafeedSubscription'

Fields

ownerId :: Maybe Text

The Amazon Web Services account ID of the account.

prefix :: Maybe Text

The prefix for the data feed files.

fault :: Maybe SpotInstanceStateFault

The fault codes for the Spot Instance request, if any.

state :: Maybe DatafeedSubscriptionState

The state of the Spot Instance data feed subscription.

bucket :: Maybe Text

The name of the Amazon S3 bucket where the Spot Instance data feed is located.

newSpotDatafeedSubscription :: SpotDatafeedSubscription

Create a value of SpotDatafeedSubscription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ownerId:SpotDatafeedSubscription', spotDatafeedSubscription_ownerId - The Amazon Web Services account ID of the account.

$sel:prefix:SpotDatafeedSubscription', spotDatafeedSubscription_prefix - The prefix for the data feed files.

$sel:fault:SpotDatafeedSubscription', spotDatafeedSubscription_fault - The fault codes for the Spot Instance request, if any.

$sel:state:SpotDatafeedSubscription', spotDatafeedSubscription_state - The state of the Spot Instance data feed subscription.

$sel:bucket:SpotDatafeedSubscription', spotDatafeedSubscription_bucket - The name of the Amazon S3 bucket where the Spot Instance data feed is located.

spotDatafeedSubscription_ownerId :: Lens' SpotDatafeedSubscription (Maybe Text)

The Amazon Web Services account ID of the account.

spotDatafeedSubscription_prefix :: Lens' SpotDatafeedSubscription (Maybe Text)

The prefix for the data feed files.

spotDatafeedSubscription_fault :: Lens' SpotDatafeedSubscription (Maybe SpotInstanceStateFault)

The fault codes for the Spot Instance request, if any.

spotDatafeedSubscription_state :: Lens' SpotDatafeedSubscription (Maybe DatafeedSubscriptionState)

The state of the Spot Instance data feed subscription.

spotDatafeedSubscription_bucket :: Lens' SpotDatafeedSubscription (Maybe Text)

The name of the Amazon S3 bucket where the Spot Instance data feed is located.