Network.AWS.EC2.Types.SpotFleetLaunchSpecification

Description

Documentation

data SpotFleetLaunchSpecification

Describes the launch specification for one or more Spot Instances. If you include On-Demand capacity in your fleet request or want to specify an EFA network device, you can't use SpotFleetLaunchSpecification; you must use LaunchTemplateConfig.

See: newSpotFleetLaunchSpecification smart constructor.

Constructors

SpotFleetLaunchSpecification'

Fields

tagSpecifications :: Maybe [SpotFleetTagSpecification]

The tags to apply during creation.

ebsOptimized :: Maybe Bool

Indicates whether the instances are optimized for EBS I/O. This optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal EBS I/O performance. This optimization isn't available with all instance types. Additional usage charges apply when using an EBS Optimized instance.

Default: false

placement :: Maybe SpotPlacement

The placement information.

userData :: Maybe Text

The Base64-encoded user data that instances use when starting up.

instanceType :: Maybe InstanceType

The instance type.

ramdiskId :: Maybe Text

The ID of the RAM disk. Some kernels require additional drivers at launch. Check the kernel requirements for information about whether you need to specify a RAM disk. To find kernel requirements, refer to the Amazon Web Services Resource Center and search for the kernel ID.

addressingType :: Maybe Text

Deprecated.

spotPrice :: Maybe Text

The maximum price per unit hour that you are willing to pay for a Spot Instance. If this value is not specified, the default is the Spot price specified for the fleet. To determine the Spot price per unit hour, divide the Spot price by the value of WeightedCapacity.

imageId :: Maybe Text

The ID of the AMI.

securityGroups :: Maybe [GroupIdentifier]

One or more security groups. When requesting instances in a VPC, you must specify the IDs of the security groups. When requesting instances in EC2-Classic, you can specify the names or the IDs of the security groups.

iamInstanceProfile :: Maybe IamInstanceProfileSpecification

The IAM instance profile.

monitoring :: Maybe SpotFleetMonitoring

Enable or disable monitoring for the instances.

blockDeviceMappings :: Maybe [BlockDeviceMapping]

One or more block devices that are mapped to the Spot Instances. You can't specify both a snapshot ID and an encryption value. This is because only blank volumes can be encrypted on creation. If a snapshot is the basis for a volume, it is not blank and its encryption status is used for the volume encryption status.

subnetId :: Maybe Text

The IDs of the subnets in which to launch the instances. To specify multiple subnets, separate them using commas; for example, "subnet-1234abcdeexample1, subnet-0987cdef6example2".

kernelId :: Maybe Text

The ID of the kernel.

networkInterfaces :: Maybe [InstanceNetworkInterfaceSpecification]

One or more network interfaces. If you specify a network interface, you must specify subnet IDs and security group IDs using the network interface.

SpotFleetLaunchSpecification currently does not support Elastic Fabric Adapter (EFA). To specify an EFA, you must use LaunchTemplateConfig.

keyName :: Maybe Text

The name of the key pair.

weightedCapacity :: Maybe Double

The number of units provided by the specified instance type. These are the same units that you chose to set the target capacity in terms of instances, or a performance characteristic such as vCPUs, memory, or I/O.

If the target capacity divided by this value is not a whole number, Amazon EC2 rounds the number of instances to the next whole number. If this value is not specified, the default is 1.

newSpotFleetLaunchSpecification :: SpotFleetLaunchSpecification

Create a value of SpotFleetLaunchSpecification with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tagSpecifications:SpotFleetLaunchSpecification', spotFleetLaunchSpecification_tagSpecifications - The tags to apply during creation.

$sel:ebsOptimized:SpotFleetLaunchSpecification', spotFleetLaunchSpecification_ebsOptimized - Indicates whether the instances are optimized for EBS I/O. This optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal EBS I/O performance. This optimization isn't available with all instance types. Additional usage charges apply when using an EBS Optimized instance.

Default: false

$sel:placement:SpotFleetLaunchSpecification', spotFleetLaunchSpecification_placement - The placement information.

$sel:userData:SpotFleetLaunchSpecification', spotFleetLaunchSpecification_userData - The Base64-encoded user data that instances use when starting up.

$sel:instanceType:SpotFleetLaunchSpecification', spotFleetLaunchSpecification_instanceType - The instance type.

$sel:ramdiskId:SpotFleetLaunchSpecification', spotFleetLaunchSpecification_ramdiskId - The ID of the RAM disk. Some kernels require additional drivers at launch. Check the kernel requirements for information about whether you need to specify a RAM disk. To find kernel requirements, refer to the Amazon Web Services Resource Center and search for the kernel ID.

$sel:addressingType:SpotFleetLaunchSpecification', spotFleetLaunchSpecification_addressingType - Deprecated.

$sel:spotPrice:SpotFleetLaunchSpecification', spotFleetLaunchSpecification_spotPrice - The maximum price per unit hour that you are willing to pay for a Spot Instance. If this value is not specified, the default is the Spot price specified for the fleet. To determine the Spot price per unit hour, divide the Spot price by the value of WeightedCapacity.

$sel:imageId:SpotFleetLaunchSpecification', spotFleetLaunchSpecification_imageId - The ID of the AMI.

$sel:securityGroups:SpotFleetLaunchSpecification', spotFleetLaunchSpecification_securityGroups - One or more security groups. When requesting instances in a VPC, you must specify the IDs of the security groups. When requesting instances in EC2-Classic, you can specify the names or the IDs of the security groups.

$sel:iamInstanceProfile:SpotFleetLaunchSpecification', spotFleetLaunchSpecification_iamInstanceProfile - The IAM instance profile.

$sel:monitoring:SpotFleetLaunchSpecification', spotFleetLaunchSpecification_monitoring - Enable or disable monitoring for the instances.

$sel:blockDeviceMappings:SpotFleetLaunchSpecification', spotFleetLaunchSpecification_blockDeviceMappings - One or more block devices that are mapped to the Spot Instances. You can't specify both a snapshot ID and an encryption value. This is because only blank volumes can be encrypted on creation. If a snapshot is the basis for a volume, it is not blank and its encryption status is used for the volume encryption status.

$sel:subnetId:SpotFleetLaunchSpecification', spotFleetLaunchSpecification_subnetId - The IDs of the subnets in which to launch the instances. To specify multiple subnets, separate them using commas; for example, "subnet-1234abcdeexample1, subnet-0987cdef6example2".

$sel:kernelId:SpotFleetLaunchSpecification', spotFleetLaunchSpecification_kernelId - The ID of the kernel.

$sel:networkInterfaces:SpotFleetLaunchSpecification', spotFleetLaunchSpecification_networkInterfaces - One or more network interfaces. If you specify a network interface, you must specify subnet IDs and security group IDs using the network interface.

SpotFleetLaunchSpecification currently does not support Elastic Fabric Adapter (EFA). To specify an EFA, you must use LaunchTemplateConfig.

$sel:keyName:SpotFleetLaunchSpecification', spotFleetLaunchSpecification_keyName - The name of the key pair.

$sel:weightedCapacity:SpotFleetLaunchSpecification', spotFleetLaunchSpecification_weightedCapacity - The number of units provided by the specified instance type. These are the same units that you chose to set the target capacity in terms of instances, or a performance characteristic such as vCPUs, memory, or I/O.

If the target capacity divided by this value is not a whole number, Amazon EC2 rounds the number of instances to the next whole number. If this value is not specified, the default is 1.

spotFleetLaunchSpecification_tagSpecifications :: Lens' SpotFleetLaunchSpecification (Maybe [SpotFleetTagSpecification])

The tags to apply during creation.

spotFleetLaunchSpecification_ebsOptimized :: Lens' SpotFleetLaunchSpecification (Maybe Bool)

Indicates whether the instances are optimized for EBS I/O. This optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal EBS I/O performance. This optimization isn't available with all instance types. Additional usage charges apply when using an EBS Optimized instance.

Default: false

spotFleetLaunchSpecification_placement :: Lens' SpotFleetLaunchSpecification (Maybe SpotPlacement)

The placement information.

spotFleetLaunchSpecification_userData :: Lens' SpotFleetLaunchSpecification (Maybe Text)

The Base64-encoded user data that instances use when starting up.

spotFleetLaunchSpecification_instanceType :: Lens' SpotFleetLaunchSpecification (Maybe InstanceType)

The instance type.

spotFleetLaunchSpecification_ramdiskId :: Lens' SpotFleetLaunchSpecification (Maybe Text)

The ID of the RAM disk. Some kernels require additional drivers at launch. Check the kernel requirements for information about whether you need to specify a RAM disk. To find kernel requirements, refer to the Amazon Web Services Resource Center and search for the kernel ID.

spotFleetLaunchSpecification_addressingType :: Lens' SpotFleetLaunchSpecification (Maybe Text)

Deprecated.

spotFleetLaunchSpecification_spotPrice :: Lens' SpotFleetLaunchSpecification (Maybe Text)

The maximum price per unit hour that you are willing to pay for a Spot Instance. If this value is not specified, the default is the Spot price specified for the fleet. To determine the Spot price per unit hour, divide the Spot price by the value of WeightedCapacity.

spotFleetLaunchSpecification_imageId :: Lens' SpotFleetLaunchSpecification (Maybe Text)

The ID of the AMI.

spotFleetLaunchSpecification_securityGroups :: Lens' SpotFleetLaunchSpecification (Maybe [GroupIdentifier])

One or more security groups. When requesting instances in a VPC, you must specify the IDs of the security groups. When requesting instances in EC2-Classic, you can specify the names or the IDs of the security groups.

spotFleetLaunchSpecification_iamInstanceProfile :: Lens' SpotFleetLaunchSpecification (Maybe IamInstanceProfileSpecification)

The IAM instance profile.

spotFleetLaunchSpecification_monitoring :: Lens' SpotFleetLaunchSpecification (Maybe SpotFleetMonitoring)

Enable or disable monitoring for the instances.

spotFleetLaunchSpecification_blockDeviceMappings :: Lens' SpotFleetLaunchSpecification (Maybe [BlockDeviceMapping])

One or more block devices that are mapped to the Spot Instances. You can't specify both a snapshot ID and an encryption value. This is because only blank volumes can be encrypted on creation. If a snapshot is the basis for a volume, it is not blank and its encryption status is used for the volume encryption status.

spotFleetLaunchSpecification_subnetId :: Lens' SpotFleetLaunchSpecification (Maybe Text)

The IDs of the subnets in which to launch the instances. To specify multiple subnets, separate them using commas; for example, "subnet-1234abcdeexample1, subnet-0987cdef6example2".

spotFleetLaunchSpecification_kernelId :: Lens' SpotFleetLaunchSpecification (Maybe Text)

The ID of the kernel.

spotFleetLaunchSpecification_networkInterfaces :: Lens' SpotFleetLaunchSpecification (Maybe [InstanceNetworkInterfaceSpecification])

One or more network interfaces. If you specify a network interface, you must specify subnet IDs and security group IDs using the network interface.

SpotFleetLaunchSpecification currently does not support Elastic Fabric Adapter (EFA). To specify an EFA, you must use LaunchTemplateConfig.

spotFleetLaunchSpecification_keyName :: Lens' SpotFleetLaunchSpecification (Maybe Text)

The name of the key pair.

spotFleetLaunchSpecification_weightedCapacity :: Lens' SpotFleetLaunchSpecification (Maybe Double)

The number of units provided by the specified instance type. These are the same units that you chose to set the target capacity in terms of instances, or a performance characteristic such as vCPUs, memory, or I/O.

If the target capacity divided by this value is not a whole number, Amazon EC2 rounds the number of instances to the next whole number. If this value is not specified, the default is 1.