Network.AWS.EC2.Types.SpotFleetRequestConfig

Description

Documentation

data SpotFleetRequestConfig

Describes a Spot Fleet request.

See: newSpotFleetRequestConfig smart constructor.

Constructors

SpotFleetRequestConfig'

Fields

spotFleetRequestState :: Maybe BatchState

The state of the Spot Fleet request.

createTime :: Maybe ISO8601

The creation date and time of the request.

tags :: Maybe [Tag]

The tags for a Spot Fleet resource.

activityStatus :: Maybe ActivityStatus

The progress of the Spot Fleet request. If there is an error, the status is error. After all requests are placed, the status is pending_fulfillment. If the size of the fleet is equal to or greater than its target capacity, the status is fulfilled. If the size of the fleet is decreased, the status is pending_termination while Spot Instances are terminating.

spotFleetRequestConfig :: Maybe SpotFleetRequestConfigData

The configuration of the Spot Fleet request.

spotFleetRequestId :: Maybe Text

The ID of the Spot Fleet request.

newSpotFleetRequestConfig :: SpotFleetRequestConfig

Create a value of SpotFleetRequestConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:spotFleetRequestState:SpotFleetRequestConfig', spotFleetRequestConfig_spotFleetRequestState - The state of the Spot Fleet request.

$sel:createTime:SpotFleetRequestConfig', spotFleetRequestConfig_createTime - The creation date and time of the request.

$sel:tags:SpotFleetRequestConfig', spotFleetRequestConfig_tags - The tags for a Spot Fleet resource.

$sel:activityStatus:SpotFleetRequestConfig', spotFleetRequestConfig_activityStatus - The progress of the Spot Fleet request. If there is an error, the status is error. After all requests are placed, the status is pending_fulfillment. If the size of the fleet is equal to or greater than its target capacity, the status is fulfilled. If the size of the fleet is decreased, the status is pending_termination while Spot Instances are terminating.

$sel:spotFleetRequestConfig:SpotFleetRequestConfig', spotFleetRequestConfig_spotFleetRequestConfig - The configuration of the Spot Fleet request.

$sel:spotFleetRequestId:SpotFleetRequestConfig', spotFleetRequestConfig_spotFleetRequestId - The ID of the Spot Fleet request.

spotFleetRequestConfig_spotFleetRequestState :: Lens' SpotFleetRequestConfig (Maybe BatchState)

The state of the Spot Fleet request.

spotFleetRequestConfig_createTime :: Lens' SpotFleetRequestConfig (Maybe UTCTime)

The creation date and time of the request.

spotFleetRequestConfig_tags :: Lens' SpotFleetRequestConfig (Maybe [Tag])

The tags for a Spot Fleet resource.

spotFleetRequestConfig_activityStatus :: Lens' SpotFleetRequestConfig (Maybe ActivityStatus)

The progress of the Spot Fleet request. If there is an error, the status is error. After all requests are placed, the status is pending_fulfillment. If the size of the fleet is equal to or greater than its target capacity, the status is fulfilled. If the size of the fleet is decreased, the status is pending_termination while Spot Instances are terminating.

spotFleetRequestConfig_spotFleetRequestConfig :: Lens' SpotFleetRequestConfig (Maybe SpotFleetRequestConfigData)

The configuration of the Spot Fleet request.

spotFleetRequestConfig_spotFleetRequestId :: Lens' SpotFleetRequestConfig (Maybe Text)

The ID of the Spot Fleet request.