Network.AWS.EC2.Types.SpotFleetRequestConfigData

Description

Documentation

data SpotFleetRequestConfigData

Describes the configuration of a Spot Fleet request.

See: newSpotFleetRequestConfigData smart constructor.

Constructors

SpotFleetRequestConfigData'

Fields

context :: Maybe Text

Reserved.

clientToken :: Maybe Text

A unique, case-sensitive identifier that you provide to ensure the idempotency of your listings. This helps to avoid duplicate listings. For more information, see Ensuring Idempotency.

instanceInterruptionBehavior :: Maybe InstanceInterruptionBehavior

The behavior when a Spot Instance is interrupted. The default is terminate.

onDemandMaxTotalPrice :: Maybe Text

The maximum amount per hour for On-Demand Instances that you're willing to pay. You can use the onDemandMaxTotalPrice parameter, the spotMaxTotalPrice parameter, or both parameters to ensure that your fleet cost does not exceed your budget. If you set a maximum price per hour for the On-Demand Instances and Spot Instances in your request, Spot Fleet will launch instances until it reaches the maximum amount you're willing to pay. When the maximum amount you're willing to pay is reached, the fleet stops launching instances even if it hasn’t met the target capacity.

spotPrice :: Maybe Text

The maximum price per unit hour that you are willing to pay for a Spot Instance. The default is the On-Demand price.

spotMaintenanceStrategies :: Maybe SpotMaintenanceStrategies

The strategies for managing your Spot Instances that are at an elevated risk of being interrupted.

loadBalancersConfig :: Maybe LoadBalancersConfig

One or more Classic Load Balancers and target groups to attach to the Spot Fleet request. Spot Fleet registers the running Spot Instances with the specified Classic Load Balancers and target groups.

With Network Load Balancers, Spot Fleet cannot register instances that have the following instance types: C1, CC1, CC2, CG1, CG2, CR1, CS1, G1, G2, HI1, HS1, M1, M2, M3, and T1.

excessCapacityTerminationPolicy :: Maybe ExcessCapacityTerminationPolicy

Indicates whether running Spot Instances should be terminated if you decrease the target capacity of the Spot Fleet request below the current size of the Spot Fleet.

onDemandTargetCapacity :: Maybe Int

The number of On-Demand units to request. You can choose to set the target capacity in terms of instances or a performance characteristic that is important to your application workload, such as vCPUs, memory, or I/O. If the request type is maintain, you can specify a target capacity of 0 and add capacity later.

launchTemplateConfigs :: Maybe [LaunchTemplateConfig]

The launch template and overrides. If you specify LaunchTemplateConfigs, you can't specify LaunchSpecifications. If you include On-Demand capacity in your request, you must use LaunchTemplateConfigs.

tagSpecifications :: Maybe [TagSpecification]

The key-value pair for tagging the Spot Fleet request on creation. The value for ResourceType must be spot-fleet-request, otherwise the Spot Fleet request fails. To tag instances at launch, specify the tags in the launch template (valid only if you use LaunchTemplateConfigs) or in the SpotFleetTagSpecification (valid only if you use LaunchSpecifications). For information about tagging after launch, see Tagging Your Resources.

validUntil :: Maybe ISO8601

The end date and time of the request, in UTC format (YYYY-MM-DDTHH:MM:SSZ). After the end date and time, no new Spot Instance requests are placed or able to fulfill the request. If no value is specified, the Spot Fleet request remains until you cancel it.

terminateInstancesWithExpiration :: Maybe Bool

Indicates whether running Spot Instances are terminated when the Spot Fleet request expires.

onDemandAllocationStrategy :: Maybe OnDemandAllocationStrategy

The order of the launch template overrides to use in fulfilling On-Demand capacity. If you specify lowestPrice, Spot Fleet uses price to determine the order, launching the lowest price first. If you specify prioritized, Spot Fleet uses the priority that you assign to each Spot Fleet launch template override, launching the highest priority first. If you do not specify a value, Spot Fleet defaults to lowestPrice.

instancePoolsToUseCount :: Maybe Int

The number of Spot pools across which to allocate your target Spot capacity. Valid only when Spot AllocationStrategy is set to lowest-price. Spot Fleet selects the cheapest Spot pools and evenly allocates your target Spot capacity across the number of Spot pools that you specify.

Note that Spot Fleet attempts to draw Spot Instances from the number of pools that you specify on a best effort basis. If a pool runs out of Spot capacity before fulfilling your target capacity, Spot Fleet will continue to fulfill your request by drawing from the next cheapest pool. To ensure that your target capacity is met, you might receive Spot Instances from more than the number of pools that you specified. Similarly, if most of the pools have no Spot capacity, you might receive your full target capacity from fewer than the number of pools that you specified.

fulfilledCapacity :: Maybe Double

The number of units fulfilled by this request compared to the set target capacity. You cannot set this value.

type' :: Maybe FleetType

The type of request. Indicates whether the Spot Fleet only requests the target capacity or also attempts to maintain it. When this value is request, the Spot Fleet only places the required requests. It does not attempt to replenish Spot Instances if capacity is diminished, nor does it submit requests in alternative Spot pools if capacity is not available. When this value is maintain, the Spot Fleet maintains the target capacity. The Spot Fleet places the required requests to meet capacity and automatically replenishes any interrupted instances. Default: maintain. instant is listed but is not used by Spot Fleet.

validFrom :: Maybe ISO8601

The start date and time of the request, in UTC format (YYYY-MM-DDTHH:MM:SSZ). By default, Amazon EC2 starts fulfilling the request immediately.

replaceUnhealthyInstances :: Maybe Bool

Indicates whether Spot Fleet should replace unhealthy instances.

launchSpecifications :: Maybe [SpotFleetLaunchSpecification]

The launch specifications for the Spot Fleet request. If you specify LaunchSpecifications, you can't specify LaunchTemplateConfigs. If you include On-Demand capacity in your request, you must use LaunchTemplateConfigs.

onDemandFulfilledCapacity :: Maybe Double

The number of On-Demand units fulfilled by this request compared to the set target On-Demand capacity.

spotMaxTotalPrice :: Maybe Text

The maximum amount per hour for Spot Instances that you're willing to pay. You can use the spotdMaxTotalPrice parameter, the onDemandMaxTotalPrice parameter, or both parameters to ensure that your fleet cost does not exceed your budget. If you set a maximum price per hour for the On-Demand Instances and Spot Instances in your request, Spot Fleet will launch instances until it reaches the maximum amount you're willing to pay. When the maximum amount you're willing to pay is reached, the fleet stops launching instances even if it hasn’t met the target capacity.

allocationStrategy :: Maybe AllocationStrategy

Indicates how to allocate the target Spot Instance capacity across the Spot Instance pools specified by the Spot Fleet request.

If the allocation strategy is lowestPrice, Spot Fleet launches instances from the Spot Instance pools with the lowest price. This is the default allocation strategy.

If the allocation strategy is diversified, Spot Fleet launches instances from all the Spot Instance pools that you specify.

If the allocation strategy is capacityOptimized (recommended), Spot Fleet launches instances from Spot Instance pools with optimal capacity for the number of instances that are launching. To give certain instance types a higher chance of launching first, use capacityOptimizedPrioritized. Set a priority for each instance type by using the Priority parameter for LaunchTemplateOverrides. You can assign the same priority to different LaunchTemplateOverrides. EC2 implements the priorities on a best-effort basis, but optimizes for capacity first. capacityOptimizedPrioritized is supported only if your Spot Fleet uses a launch template. Note that if the OnDemandAllocationStrategy is set to prioritized, the same priority is applied when fulfilling On-Demand capacity.

iamFleetRole :: Text

The Amazon Resource Name (ARN) of an Identity and Access Management (IAM) role that grants the Spot Fleet the permission to request, launch, terminate, and tag instances on your behalf. For more information, see Spot Fleet prerequisites in the Amazon EC2 User Guide for Linux Instances. Spot Fleet can terminate Spot Instances on your behalf when you cancel its Spot Fleet request using CancelSpotFleetRequests or when the Spot Fleet request expires, if you set TerminateInstancesWithExpiration.

targetCapacity :: Int

The number of units to request for the Spot Fleet. You can choose to set the target capacity in terms of instances or a performance characteristic that is important to your application workload, such as vCPUs, memory, or I/O. If the request type is maintain, you can specify a target capacity of 0 and add capacity later.

newSpotFleetRequestConfigData

Arguments

:: Text
$sel:iamFleetRole:SpotFleetRequestConfigData'

-> Int
$sel:targetCapacity:SpotFleetRequestConfigData'

-> SpotFleetRequestConfigData

Create a value of SpotFleetRequestConfigData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:context:SpotFleetRequestConfigData', spotFleetRequestConfigData_context - Reserved.

$sel:clientToken:SpotFleetRequestConfigData', spotFleetRequestConfigData_clientToken - A unique, case-sensitive identifier that you provide to ensure the idempotency of your listings. This helps to avoid duplicate listings. For more information, see Ensuring Idempotency.

$sel:instanceInterruptionBehavior:SpotFleetRequestConfigData', spotFleetRequestConfigData_instanceInterruptionBehavior - The behavior when a Spot Instance is interrupted. The default is terminate.

$sel:onDemandMaxTotalPrice:SpotFleetRequestConfigData', spotFleetRequestConfigData_onDemandMaxTotalPrice - The maximum amount per hour for On-Demand Instances that you're willing to pay. You can use the onDemandMaxTotalPrice parameter, the spotMaxTotalPrice parameter, or both parameters to ensure that your fleet cost does not exceed your budget. If you set a maximum price per hour for the On-Demand Instances and Spot Instances in your request, Spot Fleet will launch instances until it reaches the maximum amount you're willing to pay. When the maximum amount you're willing to pay is reached, the fleet stops launching instances even if it hasn’t met the target capacity.

$sel:spotPrice:SpotFleetRequestConfigData', spotFleetRequestConfigData_spotPrice - The maximum price per unit hour that you are willing to pay for a Spot Instance. The default is the On-Demand price.

$sel:spotMaintenanceStrategies:SpotFleetRequestConfigData', spotFleetRequestConfigData_spotMaintenanceStrategies - The strategies for managing your Spot Instances that are at an elevated risk of being interrupted.

$sel:loadBalancersConfig:SpotFleetRequestConfigData', spotFleetRequestConfigData_loadBalancersConfig - One or more Classic Load Balancers and target groups to attach to the Spot Fleet request. Spot Fleet registers the running Spot Instances with the specified Classic Load Balancers and target groups.

With Network Load Balancers, Spot Fleet cannot register instances that have the following instance types: C1, CC1, CC2, CG1, CG2, CR1, CS1, G1, G2, HI1, HS1, M1, M2, M3, and T1.

$sel:excessCapacityTerminationPolicy:SpotFleetRequestConfigData', spotFleetRequestConfigData_excessCapacityTerminationPolicy - Indicates whether running Spot Instances should be terminated if you decrease the target capacity of the Spot Fleet request below the current size of the Spot Fleet.

$sel:onDemandTargetCapacity:SpotFleetRequestConfigData', spotFleetRequestConfigData_onDemandTargetCapacity - The number of On-Demand units to request. You can choose to set the target capacity in terms of instances or a performance characteristic that is important to your application workload, such as vCPUs, memory, or I/O. If the request type is maintain, you can specify a target capacity of 0 and add capacity later.

$sel:launchTemplateConfigs:SpotFleetRequestConfigData', spotFleetRequestConfigData_launchTemplateConfigs - The launch template and overrides. If you specify LaunchTemplateConfigs, you can't specify LaunchSpecifications. If you include On-Demand capacity in your request, you must use LaunchTemplateConfigs.

$sel:tagSpecifications:SpotFleetRequestConfigData', spotFleetRequestConfigData_tagSpecifications - The key-value pair for tagging the Spot Fleet request on creation. The value for ResourceType must be spot-fleet-request, otherwise the Spot Fleet request fails. To tag instances at launch, specify the tags in the launch template (valid only if you use LaunchTemplateConfigs) or in the SpotFleetTagSpecification (valid only if you use LaunchSpecifications). For information about tagging after launch, see Tagging Your Resources.

$sel:validUntil:SpotFleetRequestConfigData', spotFleetRequestConfigData_validUntil - The end date and time of the request, in UTC format (YYYY-MM-DDTHH:MM:SSZ). After the end date and time, no new Spot Instance requests are placed or able to fulfill the request. If no value is specified, the Spot Fleet request remains until you cancel it.

$sel:terminateInstancesWithExpiration:SpotFleetRequestConfigData', spotFleetRequestConfigData_terminateInstancesWithExpiration - Indicates whether running Spot Instances are terminated when the Spot Fleet request expires.

$sel:onDemandAllocationStrategy:SpotFleetRequestConfigData', spotFleetRequestConfigData_onDemandAllocationStrategy - The order of the launch template overrides to use in fulfilling On-Demand capacity. If you specify lowestPrice, Spot Fleet uses price to determine the order, launching the lowest price first. If you specify prioritized, Spot Fleet uses the priority that you assign to each Spot Fleet launch template override, launching the highest priority first. If you do not specify a value, Spot Fleet defaults to lowestPrice.

$sel:instancePoolsToUseCount:SpotFleetRequestConfigData', spotFleetRequestConfigData_instancePoolsToUseCount - The number of Spot pools across which to allocate your target Spot capacity. Valid only when Spot AllocationStrategy is set to lowest-price. Spot Fleet selects the cheapest Spot pools and evenly allocates your target Spot capacity across the number of Spot pools that you specify.

Note that Spot Fleet attempts to draw Spot Instances from the number of pools that you specify on a best effort basis. If a pool runs out of Spot capacity before fulfilling your target capacity, Spot Fleet will continue to fulfill your request by drawing from the next cheapest pool. To ensure that your target capacity is met, you might receive Spot Instances from more than the number of pools that you specified. Similarly, if most of the pools have no Spot capacity, you might receive your full target capacity from fewer than the number of pools that you specified.

$sel:fulfilledCapacity:SpotFleetRequestConfigData', spotFleetRequestConfigData_fulfilledCapacity - The number of units fulfilled by this request compared to the set target capacity. You cannot set this value.

$sel:type':SpotFleetRequestConfigData', spotFleetRequestConfigData_type - The type of request. Indicates whether the Spot Fleet only requests the target capacity or also attempts to maintain it. When this value is request, the Spot Fleet only places the required requests. It does not attempt to replenish Spot Instances if capacity is diminished, nor does it submit requests in alternative Spot pools if capacity is not available. When this value is maintain, the Spot Fleet maintains the target capacity. The Spot Fleet places the required requests to meet capacity and automatically replenishes any interrupted instances. Default: maintain. instant is listed but is not used by Spot Fleet.

$sel:validFrom:SpotFleetRequestConfigData', spotFleetRequestConfigData_validFrom - The start date and time of the request, in UTC format (YYYY-MM-DDTHH:MM:SSZ). By default, Amazon EC2 starts fulfilling the request immediately.

$sel:replaceUnhealthyInstances:SpotFleetRequestConfigData', spotFleetRequestConfigData_replaceUnhealthyInstances - Indicates whether Spot Fleet should replace unhealthy instances.

$sel:launchSpecifications:SpotFleetRequestConfigData', spotFleetRequestConfigData_launchSpecifications - The launch specifications for the Spot Fleet request. If you specify LaunchSpecifications, you can't specify LaunchTemplateConfigs. If you include On-Demand capacity in your request, you must use LaunchTemplateConfigs.

$sel:onDemandFulfilledCapacity:SpotFleetRequestConfigData', spotFleetRequestConfigData_onDemandFulfilledCapacity - The number of On-Demand units fulfilled by this request compared to the set target On-Demand capacity.

$sel:spotMaxTotalPrice:SpotFleetRequestConfigData', spotFleetRequestConfigData_spotMaxTotalPrice - The maximum amount per hour for Spot Instances that you're willing to pay. You can use the spotdMaxTotalPrice parameter, the onDemandMaxTotalPrice parameter, or both parameters to ensure that your fleet cost does not exceed your budget. If you set a maximum price per hour for the On-Demand Instances and Spot Instances in your request, Spot Fleet will launch instances until it reaches the maximum amount you're willing to pay. When the maximum amount you're willing to pay is reached, the fleet stops launching instances even if it hasn’t met the target capacity.

$sel:allocationStrategy:SpotFleetRequestConfigData', spotFleetRequestConfigData_allocationStrategy - Indicates how to allocate the target Spot Instance capacity across the Spot Instance pools specified by the Spot Fleet request.

If the allocation strategy is lowestPrice, Spot Fleet launches instances from the Spot Instance pools with the lowest price. This is the default allocation strategy.

If the allocation strategy is diversified, Spot Fleet launches instances from all the Spot Instance pools that you specify.

If the allocation strategy is capacityOptimized (recommended), Spot Fleet launches instances from Spot Instance pools with optimal capacity for the number of instances that are launching. To give certain instance types a higher chance of launching first, use capacityOptimizedPrioritized. Set a priority for each instance type by using the Priority parameter for LaunchTemplateOverrides. You can assign the same priority to different LaunchTemplateOverrides. EC2 implements the priorities on a best-effort basis, but optimizes for capacity first. capacityOptimizedPrioritized is supported only if your Spot Fleet uses a launch template. Note that if the OnDemandAllocationStrategy is set to prioritized, the same priority is applied when fulfilling On-Demand capacity.

$sel:iamFleetRole:SpotFleetRequestConfigData', spotFleetRequestConfigData_iamFleetRole - The Amazon Resource Name (ARN) of an Identity and Access Management (IAM) role that grants the Spot Fleet the permission to request, launch, terminate, and tag instances on your behalf. For more information, see Spot Fleet prerequisites in the Amazon EC2 User Guide for Linux Instances. Spot Fleet can terminate Spot Instances on your behalf when you cancel its Spot Fleet request using CancelSpotFleetRequests or when the Spot Fleet request expires, if you set TerminateInstancesWithExpiration.

$sel:targetCapacity:SpotFleetRequestConfigData', spotFleetRequestConfigData_targetCapacity - The number of units to request for the Spot Fleet. You can choose to set the target capacity in terms of instances or a performance characteristic that is important to your application workload, such as vCPUs, memory, or I/O. If the request type is maintain, you can specify a target capacity of 0 and add capacity later.

spotFleetRequestConfigData_context :: Lens' SpotFleetRequestConfigData (Maybe Text)

Reserved.

spotFleetRequestConfigData_clientToken :: Lens' SpotFleetRequestConfigData (Maybe Text)

A unique, case-sensitive identifier that you provide to ensure the idempotency of your listings. This helps to avoid duplicate listings. For more information, see Ensuring Idempotency.

spotFleetRequestConfigData_instanceInterruptionBehavior :: Lens' SpotFleetRequestConfigData (Maybe InstanceInterruptionBehavior)

The behavior when a Spot Instance is interrupted. The default is terminate.

spotFleetRequestConfigData_onDemandMaxTotalPrice :: Lens' SpotFleetRequestConfigData (Maybe Text)

The maximum amount per hour for On-Demand Instances that you're willing to pay. You can use the onDemandMaxTotalPrice parameter, the spotMaxTotalPrice parameter, or both parameters to ensure that your fleet cost does not exceed your budget. If you set a maximum price per hour for the On-Demand Instances and Spot Instances in your request, Spot Fleet will launch instances until it reaches the maximum amount you're willing to pay. When the maximum amount you're willing to pay is reached, the fleet stops launching instances even if it hasn’t met the target capacity.

spotFleetRequestConfigData_spotPrice :: Lens' SpotFleetRequestConfigData (Maybe Text)

The maximum price per unit hour that you are willing to pay for a Spot Instance. The default is the On-Demand price.

spotFleetRequestConfigData_spotMaintenanceStrategies :: Lens' SpotFleetRequestConfigData (Maybe SpotMaintenanceStrategies)

The strategies for managing your Spot Instances that are at an elevated risk of being interrupted.

spotFleetRequestConfigData_loadBalancersConfig :: Lens' SpotFleetRequestConfigData (Maybe LoadBalancersConfig)

One or more Classic Load Balancers and target groups to attach to the Spot Fleet request. Spot Fleet registers the running Spot Instances with the specified Classic Load Balancers and target groups.

With Network Load Balancers, Spot Fleet cannot register instances that have the following instance types: C1, CC1, CC2, CG1, CG2, CR1, CS1, G1, G2, HI1, HS1, M1, M2, M3, and T1.

spotFleetRequestConfigData_excessCapacityTerminationPolicy :: Lens' SpotFleetRequestConfigData (Maybe ExcessCapacityTerminationPolicy)

Indicates whether running Spot Instances should be terminated if you decrease the target capacity of the Spot Fleet request below the current size of the Spot Fleet.

spotFleetRequestConfigData_onDemandTargetCapacity :: Lens' SpotFleetRequestConfigData (Maybe Int)

The number of On-Demand units to request. You can choose to set the target capacity in terms of instances or a performance characteristic that is important to your application workload, such as vCPUs, memory, or I/O. If the request type is maintain, you can specify a target capacity of 0 and add capacity later.

spotFleetRequestConfigData_launchTemplateConfigs :: Lens' SpotFleetRequestConfigData (Maybe [LaunchTemplateConfig])

The launch template and overrides. If you specify LaunchTemplateConfigs, you can't specify LaunchSpecifications. If you include On-Demand capacity in your request, you must use LaunchTemplateConfigs.

spotFleetRequestConfigData_tagSpecifications :: Lens' SpotFleetRequestConfigData (Maybe [TagSpecification])

The key-value pair for tagging the Spot Fleet request on creation. The value for ResourceType must be spot-fleet-request, otherwise the Spot Fleet request fails. To tag instances at launch, specify the tags in the launch template (valid only if you use LaunchTemplateConfigs) or in the SpotFleetTagSpecification (valid only if you use LaunchSpecifications). For information about tagging after launch, see Tagging Your Resources.

spotFleetRequestConfigData_validUntil :: Lens' SpotFleetRequestConfigData (Maybe UTCTime)

The end date and time of the request, in UTC format (YYYY-MM-DDTHH:MM:SSZ). After the end date and time, no new Spot Instance requests are placed or able to fulfill the request. If no value is specified, the Spot Fleet request remains until you cancel it.

spotFleetRequestConfigData_terminateInstancesWithExpiration :: Lens' SpotFleetRequestConfigData (Maybe Bool)

Indicates whether running Spot Instances are terminated when the Spot Fleet request expires.

spotFleetRequestConfigData_onDemandAllocationStrategy :: Lens' SpotFleetRequestConfigData (Maybe OnDemandAllocationStrategy)

The order of the launch template overrides to use in fulfilling On-Demand capacity. If you specify lowestPrice, Spot Fleet uses price to determine the order, launching the lowest price first. If you specify prioritized, Spot Fleet uses the priority that you assign to each Spot Fleet launch template override, launching the highest priority first. If you do not specify a value, Spot Fleet defaults to lowestPrice.

spotFleetRequestConfigData_instancePoolsToUseCount :: Lens' SpotFleetRequestConfigData (Maybe Int)

The number of Spot pools across which to allocate your target Spot capacity. Valid only when Spot AllocationStrategy is set to lowest-price. Spot Fleet selects the cheapest Spot pools and evenly allocates your target Spot capacity across the number of Spot pools that you specify.

Note that Spot Fleet attempts to draw Spot Instances from the number of pools that you specify on a best effort basis. If a pool runs out of Spot capacity before fulfilling your target capacity, Spot Fleet will continue to fulfill your request by drawing from the next cheapest pool. To ensure that your target capacity is met, you might receive Spot Instances from more than the number of pools that you specified. Similarly, if most of the pools have no Spot capacity, you might receive your full target capacity from fewer than the number of pools that you specified.

spotFleetRequestConfigData_fulfilledCapacity :: Lens' SpotFleetRequestConfigData (Maybe Double)

The number of units fulfilled by this request compared to the set target capacity. You cannot set this value.

spotFleetRequestConfigData_type :: Lens' SpotFleetRequestConfigData (Maybe FleetType)

The type of request. Indicates whether the Spot Fleet only requests the target capacity or also attempts to maintain it. When this value is request, the Spot Fleet only places the required requests. It does not attempt to replenish Spot Instances if capacity is diminished, nor does it submit requests in alternative Spot pools if capacity is not available. When this value is maintain, the Spot Fleet maintains the target capacity. The Spot Fleet places the required requests to meet capacity and automatically replenishes any interrupted instances. Default: maintain. instant is listed but is not used by Spot Fleet.

spotFleetRequestConfigData_validFrom :: Lens' SpotFleetRequestConfigData (Maybe UTCTime)

The start date and time of the request, in UTC format (YYYY-MM-DDTHH:MM:SSZ). By default, Amazon EC2 starts fulfilling the request immediately.

spotFleetRequestConfigData_replaceUnhealthyInstances :: Lens' SpotFleetRequestConfigData (Maybe Bool)

Indicates whether Spot Fleet should replace unhealthy instances.

spotFleetRequestConfigData_launchSpecifications :: Lens' SpotFleetRequestConfigData (Maybe [SpotFleetLaunchSpecification])

The launch specifications for the Spot Fleet request. If you specify LaunchSpecifications, you can't specify LaunchTemplateConfigs. If you include On-Demand capacity in your request, you must use LaunchTemplateConfigs.

spotFleetRequestConfigData_onDemandFulfilledCapacity :: Lens' SpotFleetRequestConfigData (Maybe Double)

The number of On-Demand units fulfilled by this request compared to the set target On-Demand capacity.

spotFleetRequestConfigData_spotMaxTotalPrice :: Lens' SpotFleetRequestConfigData (Maybe Text)

The maximum amount per hour for Spot Instances that you're willing to pay. You can use the spotdMaxTotalPrice parameter, the onDemandMaxTotalPrice parameter, or both parameters to ensure that your fleet cost does not exceed your budget. If you set a maximum price per hour for the On-Demand Instances and Spot Instances in your request, Spot Fleet will launch instances until it reaches the maximum amount you're willing to pay. When the maximum amount you're willing to pay is reached, the fleet stops launching instances even if it hasn’t met the target capacity.

spotFleetRequestConfigData_allocationStrategy :: Lens' SpotFleetRequestConfigData (Maybe AllocationStrategy)

Indicates how to allocate the target Spot Instance capacity across the Spot Instance pools specified by the Spot Fleet request.

If the allocation strategy is lowestPrice, Spot Fleet launches instances from the Spot Instance pools with the lowest price. This is the default allocation strategy.

If the allocation strategy is diversified, Spot Fleet launches instances from all the Spot Instance pools that you specify.

If the allocation strategy is capacityOptimized (recommended), Spot Fleet launches instances from Spot Instance pools with optimal capacity for the number of instances that are launching. To give certain instance types a higher chance of launching first, use capacityOptimizedPrioritized. Set a priority for each instance type by using the Priority parameter for LaunchTemplateOverrides. You can assign the same priority to different LaunchTemplateOverrides. EC2 implements the priorities on a best-effort basis, but optimizes for capacity first. capacityOptimizedPrioritized is supported only if your Spot Fleet uses a launch template. Note that if the OnDemandAllocationStrategy is set to prioritized, the same priority is applied when fulfilling On-Demand capacity.

spotFleetRequestConfigData_iamFleetRole :: Lens' SpotFleetRequestConfigData Text

The Amazon Resource Name (ARN) of an Identity and Access Management (IAM) role that grants the Spot Fleet the permission to request, launch, terminate, and tag instances on your behalf. For more information, see Spot Fleet prerequisites in the Amazon EC2 User Guide for Linux Instances. Spot Fleet can terminate Spot Instances on your behalf when you cancel its Spot Fleet request using CancelSpotFleetRequests or when the Spot Fleet request expires, if you set TerminateInstancesWithExpiration.

spotFleetRequestConfigData_targetCapacity :: Lens' SpotFleetRequestConfigData Int

The number of units to request for the Spot Fleet. You can choose to set the target capacity in terms of instances or a performance characteristic that is important to your application workload, such as vCPUs, memory, or I/O. If the request type is maintain, you can specify a target capacity of 0 and add capacity later.