Network.AWS.EC2.Types.SpotFleetTagSpecification

Description

Documentation

data SpotFleetTagSpecification

The tags for a Spot Fleet resource.

See: newSpotFleetTagSpecification smart constructor.

Constructors

SpotFleetTagSpecification'

Fields

resourceType :: Maybe ResourceType

The type of resource. Currently, the only resource type that is supported is instance. To tag the Spot Fleet request on creation, use the TagSpecifications parameter in SpotFleetRequestConfigData .

tags :: Maybe [Tag]

The tags.

newSpotFleetTagSpecification :: SpotFleetTagSpecification

Create a value of SpotFleetTagSpecification with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceType:SpotFleetTagSpecification', spotFleetTagSpecification_resourceType - The type of resource. Currently, the only resource type that is supported is instance. To tag the Spot Fleet request on creation, use the TagSpecifications parameter in SpotFleetRequestConfigData .

$sel:tags:SpotFleetTagSpecification', spotFleetTagSpecification_tags - The tags.

spotFleetTagSpecification_resourceType :: Lens' SpotFleetTagSpecification (Maybe ResourceType)

The type of resource. Currently, the only resource type that is supported is instance. To tag the Spot Fleet request on creation, use the TagSpecifications parameter in SpotFleetRequestConfigData .

spotFleetTagSpecification_tags :: Lens' SpotFleetTagSpecification (Maybe [Tag])

The tags.