Network.AWS.EC2.Types.SpotInstanceRequest

Description

Documentation

data SpotInstanceRequest

Describes a Spot Instance request.

See: newSpotInstanceRequest smart constructor.

Constructors

SpotInstanceRequest'

Fields

actualBlockHourlyPrice :: Maybe Text

Deprecated.

status :: Maybe SpotInstanceStatus

The status code and status message describing the Spot Instance request.

instanceId :: Maybe Text

The instance ID, if an instance has been launched to fulfill the Spot Instance request.

launchedAvailabilityZone :: Maybe Text

The Availability Zone in which the request is launched.

validFrom :: Maybe ISO8601

The start date of the request, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ). The request becomes active at this date and time.

spotPrice :: Maybe Text

The maximum price per hour that you are willing to pay for a Spot Instance.

launchGroup :: Maybe Text

The instance launch group. Launch groups are Spot Instances that launch together and terminate together.

blockDurationMinutes :: Maybe Int

Deprecated.

fault :: Maybe SpotInstanceStateFault

The fault codes for the Spot Instance request, if any.

instanceInterruptionBehavior :: Maybe InstanceInterruptionBehavior

The behavior when a Spot Instance is interrupted.

state :: Maybe SpotInstanceState

The state of the Spot Instance request. Spot status information helps track your Spot Instance requests. For more information, see Spot status in the Amazon EC2 User Guide for Linux Instances.

validUntil :: Maybe ISO8601

The end date of the request, in UTC format (YYYY-MM-DDTHH:MM:SSZ).

For a persistent request, the request remains active until the validUntil date and time is reached. Otherwise, the request remains active until you cancel it.

For a one-time request, the request remains active until all instances launch, the request is canceled, or the validUntil date and time is reached. By default, the request is valid for 7 days from the date the request was created.

createTime :: Maybe ISO8601

The date and time when the Spot Instance request was created, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ).

tags :: Maybe [Tag]

Any tags assigned to the resource.

launchSpecification :: Maybe LaunchSpecification

Additional information for launching instances.

availabilityZoneGroup :: Maybe Text

The Availability Zone group. If you specify the same Availability Zone group for all Spot Instance requests, all Spot Instances are launched in the same Availability Zone.

type' :: Maybe SpotInstanceType

The Spot Instance request type.

spotInstanceRequestId :: Maybe Text

The ID of the Spot Instance request.

productDescription :: Maybe RIProductDescription

The product description associated with the Spot Instance.

newSpotInstanceRequest :: SpotInstanceRequest

Create a value of SpotInstanceRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:actualBlockHourlyPrice:SpotInstanceRequest', spotInstanceRequest_actualBlockHourlyPrice - Deprecated.

$sel:status:SpotInstanceRequest', spotInstanceRequest_status - The status code and status message describing the Spot Instance request.

$sel:instanceId:SpotInstanceRequest', spotInstanceRequest_instanceId - The instance ID, if an instance has been launched to fulfill the Spot Instance request.

$sel:launchedAvailabilityZone:SpotInstanceRequest', spotInstanceRequest_launchedAvailabilityZone - The Availability Zone in which the request is launched.

$sel:validFrom:SpotInstanceRequest', spotInstanceRequest_validFrom - The start date of the request, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ). The request becomes active at this date and time.

$sel:spotPrice:SpotInstanceRequest', spotInstanceRequest_spotPrice - The maximum price per hour that you are willing to pay for a Spot Instance.

$sel:launchGroup:SpotInstanceRequest', spotInstanceRequest_launchGroup - The instance launch group. Launch groups are Spot Instances that launch together and terminate together.

$sel:blockDurationMinutes:SpotInstanceRequest', spotInstanceRequest_blockDurationMinutes - Deprecated.

$sel:fault:SpotInstanceRequest', spotInstanceRequest_fault - The fault codes for the Spot Instance request, if any.

$sel:instanceInterruptionBehavior:SpotInstanceRequest', spotInstanceRequest_instanceInterruptionBehavior - The behavior when a Spot Instance is interrupted.

$sel:state:SpotInstanceRequest', spotInstanceRequest_state - The state of the Spot Instance request. Spot status information helps track your Spot Instance requests. For more information, see Spot status in the Amazon EC2 User Guide for Linux Instances.

$sel:validUntil:SpotInstanceRequest', spotInstanceRequest_validUntil - The end date of the request, in UTC format (YYYY-MM-DDTHH:MM:SSZ).

For a persistent request, the request remains active until the validUntil date and time is reached. Otherwise, the request remains active until you cancel it.

For a one-time request, the request remains active until all instances launch, the request is canceled, or the validUntil date and time is reached. By default, the request is valid for 7 days from the date the request was created.

$sel:createTime:SpotInstanceRequest', spotInstanceRequest_createTime - The date and time when the Spot Instance request was created, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ).

$sel:tags:SpotInstanceRequest', spotInstanceRequest_tags - Any tags assigned to the resource.

$sel:launchSpecification:SpotInstanceRequest', spotInstanceRequest_launchSpecification - Additional information for launching instances.

$sel:availabilityZoneGroup:SpotInstanceRequest', spotInstanceRequest_availabilityZoneGroup - The Availability Zone group. If you specify the same Availability Zone group for all Spot Instance requests, all Spot Instances are launched in the same Availability Zone.

$sel:type':SpotInstanceRequest', spotInstanceRequest_type - The Spot Instance request type.

$sel:spotInstanceRequestId:SpotInstanceRequest', spotInstanceRequest_spotInstanceRequestId - The ID of the Spot Instance request.

$sel:productDescription:SpotInstanceRequest', spotInstanceRequest_productDescription - The product description associated with the Spot Instance.

spotInstanceRequest_actualBlockHourlyPrice :: Lens' SpotInstanceRequest (Maybe Text)

Deprecated.

spotInstanceRequest_status :: Lens' SpotInstanceRequest (Maybe SpotInstanceStatus)

The status code and status message describing the Spot Instance request.

spotInstanceRequest_instanceId :: Lens' SpotInstanceRequest (Maybe Text)

The instance ID, if an instance has been launched to fulfill the Spot Instance request.

spotInstanceRequest_launchedAvailabilityZone :: Lens' SpotInstanceRequest (Maybe Text)

The Availability Zone in which the request is launched.

spotInstanceRequest_validFrom :: Lens' SpotInstanceRequest (Maybe UTCTime)

The start date of the request, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ). The request becomes active at this date and time.

spotInstanceRequest_spotPrice :: Lens' SpotInstanceRequest (Maybe Text)

The maximum price per hour that you are willing to pay for a Spot Instance.

spotInstanceRequest_launchGroup :: Lens' SpotInstanceRequest (Maybe Text)

The instance launch group. Launch groups are Spot Instances that launch together and terminate together.

spotInstanceRequest_blockDurationMinutes :: Lens' SpotInstanceRequest (Maybe Int)

Deprecated.

spotInstanceRequest_fault :: Lens' SpotInstanceRequest (Maybe SpotInstanceStateFault)

The fault codes for the Spot Instance request, if any.

spotInstanceRequest_instanceInterruptionBehavior :: Lens' SpotInstanceRequest (Maybe InstanceInterruptionBehavior)

The behavior when a Spot Instance is interrupted.

spotInstanceRequest_state :: Lens' SpotInstanceRequest (Maybe SpotInstanceState)

The state of the Spot Instance request. Spot status information helps track your Spot Instance requests. For more information, see Spot status in the Amazon EC2 User Guide for Linux Instances.

spotInstanceRequest_validUntil :: Lens' SpotInstanceRequest (Maybe UTCTime)

The end date of the request, in UTC format (YYYY-MM-DDTHH:MM:SSZ).

For a persistent request, the request remains active until the validUntil date and time is reached. Otherwise, the request remains active until you cancel it.

For a one-time request, the request remains active until all instances launch, the request is canceled, or the validUntil date and time is reached. By default, the request is valid for 7 days from the date the request was created.

spotInstanceRequest_createTime :: Lens' SpotInstanceRequest (Maybe UTCTime)

The date and time when the Spot Instance request was created, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ).

spotInstanceRequest_tags :: Lens' SpotInstanceRequest (Maybe [Tag])

Any tags assigned to the resource.

spotInstanceRequest_launchSpecification :: Lens' SpotInstanceRequest (Maybe LaunchSpecification)

Additional information for launching instances.

spotInstanceRequest_availabilityZoneGroup :: Lens' SpotInstanceRequest (Maybe Text)

The Availability Zone group. If you specify the same Availability Zone group for all Spot Instance requests, all Spot Instances are launched in the same Availability Zone.

spotInstanceRequest_type :: Lens' SpotInstanceRequest (Maybe SpotInstanceType)

The Spot Instance request type.

spotInstanceRequest_spotInstanceRequestId :: Lens' SpotInstanceRequest (Maybe Text)

The ID of the Spot Instance request.

spotInstanceRequest_productDescription :: Lens' SpotInstanceRequest (Maybe RIProductDescription)

The product description associated with the Spot Instance.