Network.AWS.EC2.Types.SpotInstanceStatus

Description

Documentation

data SpotInstanceStatus

Describes the status of a Spot Instance request.

See: newSpotInstanceStatus smart constructor.

Constructors

SpotInstanceStatus'

Fields

message :: Maybe Text

The description for the status code.

code :: Maybe Text

The status code. For a list of status codes, see Spot status codes in the Amazon EC2 User Guide for Linux Instances.

updateTime :: Maybe ISO8601

The date and time of the most recent status update, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ).

newSpotInstanceStatus :: SpotInstanceStatus

Create a value of SpotInstanceStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:message:SpotInstanceStatus', spotInstanceStatus_message - The description for the status code.

$sel:code:SpotInstanceStatus', spotInstanceStatus_code - The status code. For a list of status codes, see Spot status codes in the Amazon EC2 User Guide for Linux Instances.

$sel:updateTime:SpotInstanceStatus', spotInstanceStatus_updateTime - The date and time of the most recent status update, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ).

spotInstanceStatus_message :: Lens' SpotInstanceStatus (Maybe Text)

The description for the status code.

spotInstanceStatus_code :: Lens' SpotInstanceStatus (Maybe Text)

The status code. For a list of status codes, see Spot status codes in the Amazon EC2 User Guide for Linux Instances.

spotInstanceStatus_updateTime :: Lens' SpotInstanceStatus (Maybe UTCTime)

The date and time of the most recent status update, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ).