Network.AWS.EC2.Types.SpotMarketOptions

Description

Documentation

data SpotMarketOptions

The options for Spot Instances.

See: newSpotMarketOptions smart constructor.

Constructors

SpotMarketOptions'

Fields

blockDurationMinutes :: Maybe Int

Deprecated.

instanceInterruptionBehavior :: Maybe InstanceInterruptionBehavior

The behavior when a Spot Instance is interrupted. The default is terminate.

validUntil :: Maybe ISO8601

The end date of the request, in UTC format (YYYY-MM-DDTHH:MM:SSZ). Supported only for persistent requests.

For a persistent request, the request remains active until the ValidUntil date and time is reached. Otherwise, the request remains active until you cancel it.

For a one-time request, ValidUntil is not supported. The request remains active until all instances launch or you cancel the request.

spotInstanceType :: Maybe SpotInstanceType

The Spot Instance request type. For RunInstances, persistent Spot Instance requests are only supported when the instance interruption behavior is either hibernate or stop.

maxPrice :: Maybe Text

The maximum hourly price you're willing to pay for the Spot Instances. The default is the On-Demand price.

newSpotMarketOptions :: SpotMarketOptions

Create a value of SpotMarketOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:blockDurationMinutes:SpotMarketOptions', spotMarketOptions_blockDurationMinutes - Deprecated.

$sel:instanceInterruptionBehavior:SpotMarketOptions', spotMarketOptions_instanceInterruptionBehavior - The behavior when a Spot Instance is interrupted. The default is terminate.

$sel:validUntil:SpotMarketOptions', spotMarketOptions_validUntil - The end date of the request, in UTC format (YYYY-MM-DDTHH:MM:SSZ). Supported only for persistent requests.

For a persistent request, the request remains active until the ValidUntil date and time is reached. Otherwise, the request remains active until you cancel it.

For a one-time request, ValidUntil is not supported. The request remains active until all instances launch or you cancel the request.

$sel:spotInstanceType:SpotMarketOptions', spotMarketOptions_spotInstanceType - The Spot Instance request type. For RunInstances, persistent Spot Instance requests are only supported when the instance interruption behavior is either hibernate or stop.

$sel:maxPrice:SpotMarketOptions', spotMarketOptions_maxPrice - The maximum hourly price you're willing to pay for the Spot Instances. The default is the On-Demand price.

spotMarketOptions_blockDurationMinutes :: Lens' SpotMarketOptions (Maybe Int)

Deprecated.

spotMarketOptions_instanceInterruptionBehavior :: Lens' SpotMarketOptions (Maybe InstanceInterruptionBehavior)

The behavior when a Spot Instance is interrupted. The default is terminate.

spotMarketOptions_validUntil :: Lens' SpotMarketOptions (Maybe UTCTime)

The end date of the request, in UTC format (YYYY-MM-DDTHH:MM:SSZ). Supported only for persistent requests.

For a persistent request, the request remains active until the ValidUntil date and time is reached. Otherwise, the request remains active until you cancel it.

For a one-time request, ValidUntil is not supported. The request remains active until all instances launch or you cancel the request.

spotMarketOptions_spotInstanceType :: Lens' SpotMarketOptions (Maybe SpotInstanceType)

The Spot Instance request type. For RunInstances, persistent Spot Instance requests are only supported when the instance interruption behavior is either hibernate or stop.

spotMarketOptions_maxPrice :: Lens' SpotMarketOptions (Maybe Text)

The maximum hourly price you're willing to pay for the Spot Instances. The default is the On-Demand price.