Network.AWS.EC2.Types.SpotOptions

Description

Documentation

data SpotOptions

Describes the configuration of Spot Instances in an EC2 Fleet.

See: newSpotOptions smart constructor.

Constructors

SpotOptions'

Fields

minTargetCapacity :: Maybe Int

The minimum target capacity for Spot Instances in the fleet. If the minimum target capacity is not reached, the fleet launches no instances.

maintenanceStrategies :: Maybe FleetSpotMaintenanceStrategies

The strategies for managing your workloads on your Spot Instances that will be interrupted. Currently only the capacity rebalance strategy is available.

instanceInterruptionBehavior :: Maybe SpotInstanceInterruptionBehavior

The behavior when a Spot Instance is interrupted. The default is terminate.

maxTotalPrice :: Maybe Text

The maximum amount per hour for Spot Instances that you're willing to pay.

allocationStrategy :: Maybe SpotAllocationStrategy

Indicates how to allocate the target Spot Instance capacity across the Spot Instance pools specified by the EC2 Fleet.

If the allocation strategy is lowest-price, EC2 Fleet launches instances from the Spot Instance pools with the lowest price. This is the default allocation strategy.

If the allocation strategy is diversified, EC2 Fleet launches instances from all of the Spot Instance pools that you specify.

If the allocation strategy is capacity-optimized (recommended), EC2 Fleet launches instances from Spot Instance pools with optimal capacity for the number of instances that are launching. To give certain instance types a higher chance of launching first, use capacity-optimized-prioritized. Set a priority for each instance type by using the Priority parameter for LaunchTemplateOverrides. You can assign the same priority to different LaunchTemplateOverrides. EC2 implements the priorities on a best-effort basis, but optimizes for capacity first. capacity-optimized-prioritized is supported only if your fleet uses a launch template. Note that if the On-Demand AllocationStrategy is set to prioritized, the same priority is applied when fulfilling On-Demand capacity.

singleInstanceType :: Maybe Bool

Indicates that the fleet uses a single instance type to launch all Spot Instances in the fleet. Supported only for fleets of type instant.

singleAvailabilityZone :: Maybe Bool

Indicates that the fleet launches all Spot Instances into a single Availability Zone. Supported only for fleets of type instant.

instancePoolsToUseCount :: Maybe Int

The number of Spot pools across which to allocate your target Spot capacity. Valid only when AllocationStrategy is set to lowest-price. EC2 Fleet selects the cheapest Spot pools and evenly allocates your target Spot capacity across the number of Spot pools that you specify.

Note that EC2 Fleet attempts to draw Spot Instances from the number of pools that you specify on a best effort basis. If a pool runs out of Spot capacity before fulfilling your target capacity, EC2 Fleet will continue to fulfill your request by drawing from the next cheapest pool. To ensure that your target capacity is met, you might receive Spot Instances from more than the number of pools that you specified. Similarly, if most of the pools have no Spot capacity, you might receive your full target capacity from fewer than the number of pools that you specified.

newSpotOptions :: SpotOptions

Create a value of SpotOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:minTargetCapacity:SpotOptions', spotOptions_minTargetCapacity - The minimum target capacity for Spot Instances in the fleet. If the minimum target capacity is not reached, the fleet launches no instances.

$sel:maintenanceStrategies:SpotOptions', spotOptions_maintenanceStrategies - The strategies for managing your workloads on your Spot Instances that will be interrupted. Currently only the capacity rebalance strategy is available.

$sel:instanceInterruptionBehavior:SpotOptions', spotOptions_instanceInterruptionBehavior - The behavior when a Spot Instance is interrupted. The default is terminate.

$sel:maxTotalPrice:SpotOptions', spotOptions_maxTotalPrice - The maximum amount per hour for Spot Instances that you're willing to pay.

$sel:allocationStrategy:SpotOptions', spotOptions_allocationStrategy - Indicates how to allocate the target Spot Instance capacity across the Spot Instance pools specified by the EC2 Fleet.

If the allocation strategy is lowest-price, EC2 Fleet launches instances from the Spot Instance pools with the lowest price. This is the default allocation strategy.

If the allocation strategy is diversified, EC2 Fleet launches instances from all of the Spot Instance pools that you specify.

If the allocation strategy is capacity-optimized (recommended), EC2 Fleet launches instances from Spot Instance pools with optimal capacity for the number of instances that are launching. To give certain instance types a higher chance of launching first, use capacity-optimized-prioritized. Set a priority for each instance type by using the Priority parameter for LaunchTemplateOverrides. You can assign the same priority to different LaunchTemplateOverrides. EC2 implements the priorities on a best-effort basis, but optimizes for capacity first. capacity-optimized-prioritized is supported only if your fleet uses a launch template. Note that if the On-Demand AllocationStrategy is set to prioritized, the same priority is applied when fulfilling On-Demand capacity.

$sel:singleInstanceType:SpotOptions', spotOptions_singleInstanceType - Indicates that the fleet uses a single instance type to launch all Spot Instances in the fleet. Supported only for fleets of type instant.

$sel:singleAvailabilityZone:SpotOptions', spotOptions_singleAvailabilityZone - Indicates that the fleet launches all Spot Instances into a single Availability Zone. Supported only for fleets of type instant.

$sel:instancePoolsToUseCount:SpotOptions', spotOptions_instancePoolsToUseCount - The number of Spot pools across which to allocate your target Spot capacity. Valid only when AllocationStrategy is set to lowest-price. EC2 Fleet selects the cheapest Spot pools and evenly allocates your target Spot capacity across the number of Spot pools that you specify.

Note that EC2 Fleet attempts to draw Spot Instances from the number of pools that you specify on a best effort basis. If a pool runs out of Spot capacity before fulfilling your target capacity, EC2 Fleet will continue to fulfill your request by drawing from the next cheapest pool. To ensure that your target capacity is met, you might receive Spot Instances from more than the number of pools that you specified. Similarly, if most of the pools have no Spot capacity, you might receive your full target capacity from fewer than the number of pools that you specified.

spotOptions_minTargetCapacity :: Lens' SpotOptions (Maybe Int)

The minimum target capacity for Spot Instances in the fleet. If the minimum target capacity is not reached, the fleet launches no instances.

spotOptions_maintenanceStrategies :: Lens' SpotOptions (Maybe FleetSpotMaintenanceStrategies)

The strategies for managing your workloads on your Spot Instances that will be interrupted. Currently only the capacity rebalance strategy is available.

spotOptions_instanceInterruptionBehavior :: Lens' SpotOptions (Maybe SpotInstanceInterruptionBehavior)

The behavior when a Spot Instance is interrupted. The default is terminate.

spotOptions_maxTotalPrice :: Lens' SpotOptions (Maybe Text)

The maximum amount per hour for Spot Instances that you're willing to pay.

spotOptions_allocationStrategy :: Lens' SpotOptions (Maybe SpotAllocationStrategy)

Indicates how to allocate the target Spot Instance capacity across the Spot Instance pools specified by the EC2 Fleet.

If the allocation strategy is lowest-price, EC2 Fleet launches instances from the Spot Instance pools with the lowest price. This is the default allocation strategy.

If the allocation strategy is diversified, EC2 Fleet launches instances from all of the Spot Instance pools that you specify.

If the allocation strategy is capacity-optimized (recommended), EC2 Fleet launches instances from Spot Instance pools with optimal capacity for the number of instances that are launching. To give certain instance types a higher chance of launching first, use capacity-optimized-prioritized. Set a priority for each instance type by using the Priority parameter for LaunchTemplateOverrides. You can assign the same priority to different LaunchTemplateOverrides. EC2 implements the priorities on a best-effort basis, but optimizes for capacity first. capacity-optimized-prioritized is supported only if your fleet uses a launch template. Note that if the On-Demand AllocationStrategy is set to prioritized, the same priority is applied when fulfilling On-Demand capacity.

spotOptions_singleInstanceType :: Lens' SpotOptions (Maybe Bool)

Indicates that the fleet uses a single instance type to launch all Spot Instances in the fleet. Supported only for fleets of type instant.

spotOptions_singleAvailabilityZone :: Lens' SpotOptions (Maybe Bool)

Indicates that the fleet launches all Spot Instances into a single Availability Zone. Supported only for fleets of type instant.

spotOptions_instancePoolsToUseCount :: Lens' SpotOptions (Maybe Int)

The number of Spot pools across which to allocate your target Spot capacity. Valid only when AllocationStrategy is set to lowest-price. EC2 Fleet selects the cheapest Spot pools and evenly allocates your target Spot capacity across the number of Spot pools that you specify.

Note that EC2 Fleet attempts to draw Spot Instances from the number of pools that you specify on a best effort basis. If a pool runs out of Spot capacity before fulfilling your target capacity, EC2 Fleet will continue to fulfill your request by drawing from the next cheapest pool. To ensure that your target capacity is met, you might receive Spot Instances from more than the number of pools that you specified. Similarly, if most of the pools have no Spot capacity, you might receive your full target capacity from fewer than the number of pools that you specified.