Network.AWS.EC2.Types.SpotOptionsRequest

Description

Documentation

data SpotOptionsRequest

Describes the configuration of Spot Instances in an EC2 Fleet request.

See: newSpotOptionsRequest smart constructor.

Constructors

SpotOptionsRequest'

Fields

instanceInterruptionBehavior :: Maybe SpotInstanceInterruptionBehavior

The behavior when a Spot Instance is interrupted. The default is terminate.

singleAvailabilityZone :: Maybe Bool

Indicates that the fleet launches all Spot Instances into a single Availability Zone. Supported only for fleets of type instant.

maxTotalPrice :: Maybe Text

The maximum amount per hour for Spot Instances that you're willing to pay.

minTargetCapacity :: Maybe Int

The minimum target capacity for Spot Instances in the fleet. If the minimum target capacity is not reached, the fleet launches no instances.

instancePoolsToUseCount :: Maybe Int

The number of Spot pools across which to allocate your target Spot capacity. Valid only when Spot AllocationStrategy is set to lowest-price. EC2 Fleet selects the cheapest Spot pools and evenly allocates your target Spot capacity across the number of Spot pools that you specify.

Note that EC2 Fleet attempts to draw Spot Instances from the number of pools that you specify on a best effort basis. If a pool runs out of Spot capacity before fulfilling your target capacity, EC2 Fleet will continue to fulfill your request by drawing from the next cheapest pool. To ensure that your target capacity is met, you might receive Spot Instances from more than the number of pools that you specified. Similarly, if most of the pools have no Spot capacity, you might receive your full target capacity from fewer than the number of pools that you specified.

maintenanceStrategies :: Maybe FleetSpotMaintenanceStrategiesRequest

The strategies for managing your Spot Instances that are at an elevated risk of being interrupted.

singleInstanceType :: Maybe Bool

Indicates that the fleet uses a single instance type to launch all Spot Instances in the fleet. Supported only for fleets of type instant.

allocationStrategy :: Maybe SpotAllocationStrategy

Indicates how to allocate the target Spot Instance capacity across the Spot Instance pools specified by the EC2 Fleet.

If the allocation strategy is lowest-price, EC2 Fleet launches instances from the Spot Instance pools with the lowest price. This is the default allocation strategy.

If the allocation strategy is diversified, EC2 Fleet launches instances from all of the Spot Instance pools that you specify.

If the allocation strategy is capacity-optimized (recommended), EC2 Fleet launches instances from Spot Instance pools with optimal capacity for the number of instances that are launching. To give certain instance types a higher chance of launching first, use capacity-optimized-prioritized. Set a priority for each instance type by using the Priority parameter for LaunchTemplateOverrides. You can assign the same priority to different LaunchTemplateOverrides. EC2 implements the priorities on a best-effort basis, but optimizes for capacity first. capacity-optimized-prioritized is supported only if your fleet uses a launch template. Note that if the On-Demand AllocationStrategy is set to prioritized, the same priority is applied when fulfilling On-Demand capacity.

newSpotOptionsRequest :: SpotOptionsRequest

Create a value of SpotOptionsRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceInterruptionBehavior:SpotOptionsRequest', spotOptionsRequest_instanceInterruptionBehavior - The behavior when a Spot Instance is interrupted. The default is terminate.

$sel:singleAvailabilityZone:SpotOptionsRequest', spotOptionsRequest_singleAvailabilityZone - Indicates that the fleet launches all Spot Instances into a single Availability Zone. Supported only for fleets of type instant.

$sel:maxTotalPrice:SpotOptionsRequest', spotOptionsRequest_maxTotalPrice - The maximum amount per hour for Spot Instances that you're willing to pay.

$sel:minTargetCapacity:SpotOptionsRequest', spotOptionsRequest_minTargetCapacity - The minimum target capacity for Spot Instances in the fleet. If the minimum target capacity is not reached, the fleet launches no instances.

$sel:instancePoolsToUseCount:SpotOptionsRequest', spotOptionsRequest_instancePoolsToUseCount - The number of Spot pools across which to allocate your target Spot capacity. Valid only when Spot AllocationStrategy is set to lowest-price. EC2 Fleet selects the cheapest Spot pools and evenly allocates your target Spot capacity across the number of Spot pools that you specify.

Note that EC2 Fleet attempts to draw Spot Instances from the number of pools that you specify on a best effort basis. If a pool runs out of Spot capacity before fulfilling your target capacity, EC2 Fleet will continue to fulfill your request by drawing from the next cheapest pool. To ensure that your target capacity is met, you might receive Spot Instances from more than the number of pools that you specified. Similarly, if most of the pools have no Spot capacity, you might receive your full target capacity from fewer than the number of pools that you specified.

$sel:maintenanceStrategies:SpotOptionsRequest', spotOptionsRequest_maintenanceStrategies - The strategies for managing your Spot Instances that are at an elevated risk of being interrupted.

$sel:singleInstanceType:SpotOptionsRequest', spotOptionsRequest_singleInstanceType - Indicates that the fleet uses a single instance type to launch all Spot Instances in the fleet. Supported only for fleets of type instant.

$sel:allocationStrategy:SpotOptionsRequest', spotOptionsRequest_allocationStrategy - Indicates how to allocate the target Spot Instance capacity across the Spot Instance pools specified by the EC2 Fleet.

If the allocation strategy is lowest-price, EC2 Fleet launches instances from the Spot Instance pools with the lowest price. This is the default allocation strategy.

If the allocation strategy is diversified, EC2 Fleet launches instances from all of the Spot Instance pools that you specify.

If the allocation strategy is capacity-optimized (recommended), EC2 Fleet launches instances from Spot Instance pools with optimal capacity for the number of instances that are launching. To give certain instance types a higher chance of launching first, use capacity-optimized-prioritized. Set a priority for each instance type by using the Priority parameter for LaunchTemplateOverrides. You can assign the same priority to different LaunchTemplateOverrides. EC2 implements the priorities on a best-effort basis, but optimizes for capacity first. capacity-optimized-prioritized is supported only if your fleet uses a launch template. Note that if the On-Demand AllocationStrategy is set to prioritized, the same priority is applied when fulfilling On-Demand capacity.

spotOptionsRequest_instanceInterruptionBehavior :: Lens' SpotOptionsRequest (Maybe SpotInstanceInterruptionBehavior)

The behavior when a Spot Instance is interrupted. The default is terminate.

spotOptionsRequest_singleAvailabilityZone :: Lens' SpotOptionsRequest (Maybe Bool)

Indicates that the fleet launches all Spot Instances into a single Availability Zone. Supported only for fleets of type instant.

spotOptionsRequest_maxTotalPrice :: Lens' SpotOptionsRequest (Maybe Text)

The maximum amount per hour for Spot Instances that you're willing to pay.

spotOptionsRequest_minTargetCapacity :: Lens' SpotOptionsRequest (Maybe Int)

The minimum target capacity for Spot Instances in the fleet. If the minimum target capacity is not reached, the fleet launches no instances.

spotOptionsRequest_instancePoolsToUseCount :: Lens' SpotOptionsRequest (Maybe Int)

The number of Spot pools across which to allocate your target Spot capacity. Valid only when Spot AllocationStrategy is set to lowest-price. EC2 Fleet selects the cheapest Spot pools and evenly allocates your target Spot capacity across the number of Spot pools that you specify.

Note that EC2 Fleet attempts to draw Spot Instances from the number of pools that you specify on a best effort basis. If a pool runs out of Spot capacity before fulfilling your target capacity, EC2 Fleet will continue to fulfill your request by drawing from the next cheapest pool. To ensure that your target capacity is met, you might receive Spot Instances from more than the number of pools that you specified. Similarly, if most of the pools have no Spot capacity, you might receive your full target capacity from fewer than the number of pools that you specified.

spotOptionsRequest_maintenanceStrategies :: Lens' SpotOptionsRequest (Maybe FleetSpotMaintenanceStrategiesRequest)

The strategies for managing your Spot Instances that are at an elevated risk of being interrupted.

spotOptionsRequest_singleInstanceType :: Lens' SpotOptionsRequest (Maybe Bool)

Indicates that the fleet uses a single instance type to launch all Spot Instances in the fleet. Supported only for fleets of type instant.

spotOptionsRequest_allocationStrategy :: Lens' SpotOptionsRequest (Maybe SpotAllocationStrategy)

Indicates how to allocate the target Spot Instance capacity across the Spot Instance pools specified by the EC2 Fleet.

If the allocation strategy is lowest-price, EC2 Fleet launches instances from the Spot Instance pools with the lowest price. This is the default allocation strategy.

If the allocation strategy is diversified, EC2 Fleet launches instances from all of the Spot Instance pools that you specify.

If the allocation strategy is capacity-optimized (recommended), EC2 Fleet launches instances from Spot Instance pools with optimal capacity for the number of instances that are launching. To give certain instance types a higher chance of launching first, use capacity-optimized-prioritized. Set a priority for each instance type by using the Priority parameter for LaunchTemplateOverrides. You can assign the same priority to different LaunchTemplateOverrides. EC2 implements the priorities on a best-effort basis, but optimizes for capacity first. capacity-optimized-prioritized is supported only if your fleet uses a launch template. Note that if the On-Demand AllocationStrategy is set to prioritized, the same priority is applied when fulfilling On-Demand capacity.