Network.AWS.EC2.Types.SpotPlacement

Description

Documentation

data SpotPlacement

Describes Spot Instance placement.

See: newSpotPlacement smart constructor.

Constructors

SpotPlacement'

Fields

groupName :: Maybe Text

The name of the placement group.

tenancy :: Maybe Tenancy

The tenancy of the instance (if the instance is running in a VPC). An instance with a tenancy of dedicated runs on single-tenant hardware. The host tenancy is not supported for Spot Instances.

availabilityZone :: Maybe Text

The Availability Zone.

Spot Fleet only

To specify multiple Availability Zones, separate them using commas; for example, "us-west-2a, us-west-2b".

newSpotPlacement :: SpotPlacement

Create a value of SpotPlacement with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groupName:SpotPlacement', spotPlacement_groupName - The name of the placement group.

$sel:tenancy:SpotPlacement', spotPlacement_tenancy - The tenancy of the instance (if the instance is running in a VPC). An instance with a tenancy of dedicated runs on single-tenant hardware. The host tenancy is not supported for Spot Instances.

$sel:availabilityZone:SpotPlacement', spotPlacement_availabilityZone - The Availability Zone.

Spot Fleet only

To specify multiple Availability Zones, separate them using commas; for example, "us-west-2a, us-west-2b".

spotPlacement_groupName :: Lens' SpotPlacement (Maybe Text)

The name of the placement group.

spotPlacement_tenancy :: Lens' SpotPlacement (Maybe Tenancy)

The tenancy of the instance (if the instance is running in a VPC). An instance with a tenancy of dedicated runs on single-tenant hardware. The host tenancy is not supported for Spot Instances.

spotPlacement_availabilityZone :: Lens' SpotPlacement (Maybe Text)

The Availability Zone.

Spot Fleet only

To specify multiple Availability Zones, separate them using commas; for example, "us-west-2a, us-west-2b".