Network.AWS.EC2.Types.SpotPrice

Description

Documentation

data SpotPrice

Describes the maximum price per hour that you are willing to pay for a Spot Instance.

See: newSpotPrice smart constructor.

Constructors

SpotPrice'

Fields

instanceType :: Maybe InstanceType

The instance type.

spotPrice :: Maybe Text

The maximum price per hour that you are willing to pay for a Spot Instance.

availabilityZone :: Maybe Text

The Availability Zone.

timestamp :: Maybe ISO8601

The date and time the request was created, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ).

productDescription :: Maybe RIProductDescription

A general description of the AMI.

newSpotPrice :: SpotPrice

Create a value of SpotPrice with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceType:SpotPrice', spotPrice_instanceType - The instance type.

$sel:spotPrice:SpotPrice', spotPrice_spotPrice - The maximum price per hour that you are willing to pay for a Spot Instance.

$sel:availabilityZone:SpotPrice', spotPrice_availabilityZone - The Availability Zone.

$sel:timestamp:SpotPrice', spotPrice_timestamp - The date and time the request was created, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ).

$sel:productDescription:SpotPrice', spotPrice_productDescription - A general description of the AMI.

spotPrice_instanceType :: Lens' SpotPrice (Maybe InstanceType)

The instance type.

spotPrice_spotPrice :: Lens' SpotPrice (Maybe Text)

The maximum price per hour that you are willing to pay for a Spot Instance.

spotPrice_availabilityZone :: Lens' SpotPrice (Maybe Text)

The Availability Zone.

spotPrice_timestamp :: Lens' SpotPrice (Maybe UTCTime)

The date and time the request was created, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ).

spotPrice_productDescription :: Lens' SpotPrice (Maybe RIProductDescription)

A general description of the AMI.