Network.AWS.EC2.Types.StaleSecurityGroup

Description

Documentation

data StaleSecurityGroup

Describes a stale security group (a security group that contains stale rules).

See: newStaleSecurityGroup smart constructor.

Constructors

StaleSecurityGroup'

Fields

vpcId :: Maybe Text

The ID of the VPC for the security group.

groupId :: Maybe Text

The ID of the security group.

groupName :: Maybe Text

The name of the security group.

staleIpPermissionsEgress :: Maybe [StaleIpPermission]

Information about the stale outbound rules in the security group.

staleIpPermissions :: Maybe [StaleIpPermission]

Information about the stale inbound rules in the security group.

description :: Maybe Text

The description of the security group.

newStaleSecurityGroup :: StaleSecurityGroup

Create a value of StaleSecurityGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpcId:StaleSecurityGroup', staleSecurityGroup_vpcId - The ID of the VPC for the security group.

$sel:groupId:StaleSecurityGroup', staleSecurityGroup_groupId - The ID of the security group.

$sel:groupName:StaleSecurityGroup', staleSecurityGroup_groupName - The name of the security group.

$sel:staleIpPermissionsEgress:StaleSecurityGroup', staleSecurityGroup_staleIpPermissionsEgress - Information about the stale outbound rules in the security group.

$sel:staleIpPermissions:StaleSecurityGroup', staleSecurityGroup_staleIpPermissions - Information about the stale inbound rules in the security group.

$sel:description:StaleSecurityGroup', staleSecurityGroup_description - The description of the security group.

staleSecurityGroup_vpcId :: Lens' StaleSecurityGroup (Maybe Text)

The ID of the VPC for the security group.

staleSecurityGroup_groupId :: Lens' StaleSecurityGroup (Maybe Text)

The ID of the security group.

staleSecurityGroup_groupName :: Lens' StaleSecurityGroup (Maybe Text)

The name of the security group.

staleSecurityGroup_staleIpPermissionsEgress :: Lens' StaleSecurityGroup (Maybe [StaleIpPermission])

Information about the stale outbound rules in the security group.

staleSecurityGroup_staleIpPermissions :: Lens' StaleSecurityGroup (Maybe [StaleIpPermission])

Information about the stale inbound rules in the security group.

staleSecurityGroup_description :: Lens' StaleSecurityGroup (Maybe Text)

The description of the security group.