Network.AWS.EC2.Types.StateReason

Description

Documentation

data StateReason

Describes a state change.

See: newStateReason smart constructor.

Constructors

StateReason'

Fields

code :: Maybe Text

The reason code for the state change.

message :: Maybe Text

The message for the state change.

Server.InsufficientInstanceCapacity: There was insufficient capacity available to satisfy the launch request.

Server.InternalError: An internal error caused the instance to terminate during launch.

Server.ScheduledStop: The instance was stopped due to a scheduled retirement.

Server.SpotInstanceShutdown: The instance was stopped because the number of Spot requests with a maximum price equal to or higher than the Spot price exceeded available capacity or because of an increase in the Spot price.

Server.SpotInstanceTermination: The instance was terminated because the number of Spot requests with a maximum price equal to or higher than the Spot price exceeded available capacity or because of an increase in the Spot price.

Client.InstanceInitiatedShutdown: The instance was shut down using the shutdown -h command from the instance.

Client.InstanceTerminated: The instance was terminated or rebooted during AMI creation.

Client.InternalError: A client error caused the instance to terminate during launch.

Client.InvalidSnapshot.NotFound: The specified snapshot was not found.

Client.UserInitiatedHibernate: Hibernation was initiated on the instance.

Client.UserInitiatedShutdown: The instance was shut down using the Amazon EC2 API.

Client.VolumeLimitExceeded: The limit on the number of EBS volumes or total storage was exceeded. Decrease usage or request an increase in your account limits.

newStateReason :: StateReason

Create a value of StateReason with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:code:StateReason', stateReason_code - The reason code for the state change.

$sel:message:StateReason', stateReason_message - The message for the state change.

Server.InsufficientInstanceCapacity: There was insufficient capacity available to satisfy the launch request.

Server.InternalError: An internal error caused the instance to terminate during launch.

Server.ScheduledStop: The instance was stopped due to a scheduled retirement.

Server.SpotInstanceShutdown: The instance was stopped because the number of Spot requests with a maximum price equal to or higher than the Spot price exceeded available capacity or because of an increase in the Spot price.

Server.SpotInstanceTermination: The instance was terminated because the number of Spot requests with a maximum price equal to or higher than the Spot price exceeded available capacity or because of an increase in the Spot price.

Client.InstanceInitiatedShutdown: The instance was shut down using the shutdown -h command from the instance.

Client.InstanceTerminated: The instance was terminated or rebooted during AMI creation.

Client.InternalError: A client error caused the instance to terminate during launch.

Client.InvalidSnapshot.NotFound: The specified snapshot was not found.

Client.UserInitiatedHibernate: Hibernation was initiated on the instance.

Client.UserInitiatedShutdown: The instance was shut down using the Amazon EC2 API.

Client.VolumeLimitExceeded: The limit on the number of EBS volumes or total storage was exceeded. Decrease usage or request an increase in your account limits.

stateReason_code :: Lens' StateReason (Maybe Text)

The reason code for the state change.

stateReason_message :: Lens' StateReason (Maybe Text)

The message for the state change.

Server.InsufficientInstanceCapacity: There was insufficient capacity available to satisfy the launch request.

Server.InternalError: An internal error caused the instance to terminate during launch.

Server.ScheduledStop: The instance was stopped due to a scheduled retirement.

Server.SpotInstanceShutdown: The instance was stopped because the number of Spot requests with a maximum price equal to or higher than the Spot price exceeded available capacity or because of an increase in the Spot price.

Server.SpotInstanceTermination: The instance was terminated because the number of Spot requests with a maximum price equal to or higher than the Spot price exceeded available capacity or because of an increase in the Spot price.

Client.InstanceInitiatedShutdown: The instance was shut down using the shutdown -h command from the instance.

Client.InstanceTerminated: The instance was terminated or rebooted during AMI creation.

Client.InternalError: A client error caused the instance to terminate during launch.

Client.InvalidSnapshot.NotFound: The specified snapshot was not found.

Client.UserInitiatedHibernate: Hibernation was initiated on the instance.

Client.UserInitiatedShutdown: The instance was shut down using the Amazon EC2 API.

Client.VolumeLimitExceeded: The limit on the number of EBS volumes or total storage was exceeded. Decrease usage or request an increase in your account limits.