Network.AWS.EC2.Types.StoreImageTaskResult

Description

Documentation

data StoreImageTaskResult

The information about the AMI store task, including the progress of the task.

See: newStoreImageTaskResult smart constructor.

Constructors

StoreImageTaskResult'

Fields

storeTaskFailureReason :: Maybe Text

If the tasks fails, the reason for the failure is returned. If the task succeeds, null is returned.

amiId :: Maybe Text

The ID of the AMI that is being stored.

progressPercentage :: Maybe Int

The progress of the task as a percentage.

bucket :: Maybe Text

The name of the Amazon S3 bucket that contains the stored AMI object.

taskStartTime :: Maybe ISO8601

The time the task started.

storeTaskState :: Maybe Text

The state of the store task (InProgress, Completed, or Failed).

s3objectKey :: Maybe Text

The name of the stored AMI object in the bucket.

newStoreImageTaskResult :: StoreImageTaskResult

Create a value of StoreImageTaskResult with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:storeTaskFailureReason:StoreImageTaskResult', storeImageTaskResult_storeTaskFailureReason - If the tasks fails, the reason for the failure is returned. If the task succeeds, null is returned.

$sel:amiId:StoreImageTaskResult', storeImageTaskResult_amiId - The ID of the AMI that is being stored.

$sel:progressPercentage:StoreImageTaskResult', storeImageTaskResult_progressPercentage - The progress of the task as a percentage.

$sel:bucket:StoreImageTaskResult', storeImageTaskResult_bucket - The name of the Amazon S3 bucket that contains the stored AMI object.

$sel:taskStartTime:StoreImageTaskResult', storeImageTaskResult_taskStartTime - The time the task started.

$sel:storeTaskState:StoreImageTaskResult', storeImageTaskResult_storeTaskState - The state of the store task (InProgress, Completed, or Failed).

$sel:s3objectKey:StoreImageTaskResult', storeImageTaskResult_s3objectKey - The name of the stored AMI object in the bucket.

storeImageTaskResult_storeTaskFailureReason :: Lens' StoreImageTaskResult (Maybe Text)

If the tasks fails, the reason for the failure is returned. If the task succeeds, null is returned.

storeImageTaskResult_amiId :: Lens' StoreImageTaskResult (Maybe Text)

The ID of the AMI that is being stored.

storeImageTaskResult_progressPercentage :: Lens' StoreImageTaskResult (Maybe Int)

The progress of the task as a percentage.

storeImageTaskResult_bucket :: Lens' StoreImageTaskResult (Maybe Text)

The name of the Amazon S3 bucket that contains the stored AMI object.

storeImageTaskResult_taskStartTime :: Lens' StoreImageTaskResult (Maybe UTCTime)

The time the task started.

storeImageTaskResult_storeTaskState :: Lens' StoreImageTaskResult (Maybe Text)

The state of the store task (InProgress, Completed, or Failed).

storeImageTaskResult_s3objectKey :: Lens' StoreImageTaskResult (Maybe Text)

The name of the stored AMI object in the bucket.