Network.AWS.EC2.Types.Subnet

Description

Documentation

data Subnet

Describes a subnet.

See: newSubnet smart constructor.

Constructors

Subnet'

Fields

customerOwnedIpv4Pool :: Maybe Text

The customer-owned IPv4 address pool associated with the subnet.

ownerId :: Maybe Text

The ID of the Amazon Web Services account that owns the subnet.

assignIpv6AddressOnCreation :: Maybe Bool

Indicates whether a network interface created in this subnet (including a network interface created by RunInstances) receives an IPv6 address.

subnetArn :: Maybe Text

The Amazon Resource Name (ARN) of the subnet.

outpostArn :: Maybe Text

The Amazon Resource Name (ARN) of the Outpost.

mapPublicIpOnLaunch :: Maybe Bool

Indicates whether instances launched in this subnet receive a public IPv4 address.

availabilityZoneId :: Maybe Text

The AZ ID of the subnet.

ipv6CidrBlockAssociationSet :: Maybe [SubnetIpv6CidrBlockAssociation]

Information about the IPv6 CIDR blocks associated with the subnet.

tags :: Maybe [Tag]

Any tags assigned to the subnet.

defaultForAz :: Maybe Bool

Indicates whether this is the default subnet for the Availability Zone.

mapCustomerOwnedIpOnLaunch :: Maybe Bool

Indicates whether a network interface created in this subnet (including a network interface created by RunInstances) receives a customer-owned IPv4 address.

availabilityZone :: Text

The Availability Zone of the subnet.

availableIpAddressCount :: Int

The number of unused private IPv4 addresses in the subnet. The IPv4 addresses for any stopped instances are considered unavailable.

cidrBlock :: Text

The IPv4 CIDR block assigned to the subnet.

state :: SubnetState

The current state of the subnet.

subnetId :: Text

The ID of the subnet.

vpcId :: Text

The ID of the VPC the subnet is in.

newSubnet

Arguments

:: Text
$sel:availabilityZone:Subnet'

-> Int
$sel:availableIpAddressCount:Subnet'

-> Text
$sel:cidrBlock:Subnet'

-> SubnetState
$sel:state:Subnet'

-> Text
$sel:subnetId:Subnet'

-> Text
$sel:vpcId:Subnet'

-> Subnet

Create a value of Subnet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:customerOwnedIpv4Pool:Subnet', subnet_customerOwnedIpv4Pool - The customer-owned IPv4 address pool associated with the subnet.

$sel:ownerId:Subnet', subnet_ownerId - The ID of the Amazon Web Services account that owns the subnet.

$sel:assignIpv6AddressOnCreation:Subnet', subnet_assignIpv6AddressOnCreation - Indicates whether a network interface created in this subnet (including a network interface created by RunInstances) receives an IPv6 address.

$sel:subnetArn:Subnet', subnet_subnetArn - The Amazon Resource Name (ARN) of the subnet.

$sel:outpostArn:Subnet', subnet_outpostArn - The Amazon Resource Name (ARN) of the Outpost.

$sel:mapPublicIpOnLaunch:Subnet', subnet_mapPublicIpOnLaunch - Indicates whether instances launched in this subnet receive a public IPv4 address.

$sel:availabilityZoneId:Subnet', subnet_availabilityZoneId - The AZ ID of the subnet.

$sel:ipv6CidrBlockAssociationSet:Subnet', subnet_ipv6CidrBlockAssociationSet - Information about the IPv6 CIDR blocks associated with the subnet.

$sel:tags:Subnet', subnet_tags - Any tags assigned to the subnet.

$sel:defaultForAz:Subnet', subnet_defaultForAz - Indicates whether this is the default subnet for the Availability Zone.

$sel:mapCustomerOwnedIpOnLaunch:Subnet', subnet_mapCustomerOwnedIpOnLaunch - Indicates whether a network interface created in this subnet (including a network interface created by RunInstances) receives a customer-owned IPv4 address.

$sel:availabilityZone:Subnet', subnet_availabilityZone - The Availability Zone of the subnet.

$sel:availableIpAddressCount:Subnet', subnet_availableIpAddressCount - The number of unused private IPv4 addresses in the subnet. The IPv4 addresses for any stopped instances are considered unavailable.

$sel:cidrBlock:Subnet', subnet_cidrBlock - The IPv4 CIDR block assigned to the subnet.

$sel:state:Subnet', subnet_state - The current state of the subnet.

$sel:subnetId:Subnet', subnet_subnetId - The ID of the subnet.

$sel:vpcId:Subnet', subnet_vpcId - The ID of the VPC the subnet is in.

subnet_customerOwnedIpv4Pool :: Lens' Subnet (Maybe Text)

The customer-owned IPv4 address pool associated with the subnet.

subnet_ownerId :: Lens' Subnet (Maybe Text)

The ID of the Amazon Web Services account that owns the subnet.

subnet_assignIpv6AddressOnCreation :: Lens' Subnet (Maybe Bool)

Indicates whether a network interface created in this subnet (including a network interface created by RunInstances) receives an IPv6 address.

subnet_subnetArn :: Lens' Subnet (Maybe Text)

The Amazon Resource Name (ARN) of the subnet.

subnet_outpostArn :: Lens' Subnet (Maybe Text)

The Amazon Resource Name (ARN) of the Outpost.

subnet_mapPublicIpOnLaunch :: Lens' Subnet (Maybe Bool)

Indicates whether instances launched in this subnet receive a public IPv4 address.

subnet_availabilityZoneId :: Lens' Subnet (Maybe Text)

The AZ ID of the subnet.

subnet_ipv6CidrBlockAssociationSet :: Lens' Subnet (Maybe [SubnetIpv6CidrBlockAssociation])

Information about the IPv6 CIDR blocks associated with the subnet.

subnet_tags :: Lens' Subnet (Maybe [Tag])

Any tags assigned to the subnet.

subnet_defaultForAz :: Lens' Subnet (Maybe Bool)

Indicates whether this is the default subnet for the Availability Zone.

subnet_mapCustomerOwnedIpOnLaunch :: Lens' Subnet (Maybe Bool)

Indicates whether a network interface created in this subnet (including a network interface created by RunInstances) receives a customer-owned IPv4 address.

subnet_availabilityZone :: Lens' Subnet Text

The Availability Zone of the subnet.

subnet_availableIpAddressCount :: Lens' Subnet Int

The number of unused private IPv4 addresses in the subnet. The IPv4 addresses for any stopped instances are considered unavailable.

subnet_cidrBlock :: Lens' Subnet Text

The IPv4 CIDR block assigned to the subnet.

subnet_state :: Lens' Subnet SubnetState

The current state of the subnet.

subnet_subnetId :: Lens' Subnet Text

The ID of the subnet.

subnet_vpcId :: Lens' Subnet Text

The ID of the VPC the subnet is in.