Network.AWS.EC2.Types.SubnetCidrReservation

Description

Documentation

data SubnetCidrReservation

Describes a subnet CIDR reservation.

See: newSubnetCidrReservation smart constructor.

Constructors

SubnetCidrReservation'

Fields

ownerId :: Maybe Text

The ID of the account that owns the subnet CIDR reservation.

reservationType :: Maybe SubnetCidrReservationType

The type of reservation.

subnetCidrReservationId :: Maybe Text

The ID of the subnet CIDR reservation.

cidr :: Maybe Text

The CIDR that has been reserved.

tags :: Maybe [Tag]

The tags assigned to the subnet CIDR reservation.

description :: Maybe Text

The description assigned to the subnet CIDR reservation.

subnetId :: Maybe Text

The ID of the subnet.

newSubnetCidrReservation :: SubnetCidrReservation

Create a value of SubnetCidrReservation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ownerId:SubnetCidrReservation', subnetCidrReservation_ownerId - The ID of the account that owns the subnet CIDR reservation.

$sel:reservationType:SubnetCidrReservation', subnetCidrReservation_reservationType - The type of reservation.

$sel:subnetCidrReservationId:SubnetCidrReservation', subnetCidrReservation_subnetCidrReservationId - The ID of the subnet CIDR reservation.

$sel:cidr:SubnetCidrReservation', subnetCidrReservation_cidr - The CIDR that has been reserved.

$sel:tags:SubnetCidrReservation', subnetCidrReservation_tags - The tags assigned to the subnet CIDR reservation.

$sel:description:SubnetCidrReservation', subnetCidrReservation_description - The description assigned to the subnet CIDR reservation.

$sel:subnetId:SubnetCidrReservation', subnetCidrReservation_subnetId - The ID of the subnet.

subnetCidrReservation_ownerId :: Lens' SubnetCidrReservation (Maybe Text)

The ID of the account that owns the subnet CIDR reservation.

subnetCidrReservation_reservationType :: Lens' SubnetCidrReservation (Maybe SubnetCidrReservationType)

The type of reservation.

subnetCidrReservation_subnetCidrReservationId :: Lens' SubnetCidrReservation (Maybe Text)

The ID of the subnet CIDR reservation.

subnetCidrReservation_cidr :: Lens' SubnetCidrReservation (Maybe Text)

The CIDR that has been reserved.

subnetCidrReservation_tags :: Lens' SubnetCidrReservation (Maybe [Tag])

The tags assigned to the subnet CIDR reservation.

subnetCidrReservation_description :: Lens' SubnetCidrReservation (Maybe Text)

The description assigned to the subnet CIDR reservation.

subnetCidrReservation_subnetId :: Lens' SubnetCidrReservation (Maybe Text)

The ID of the subnet.