Network.AWS.EC2.Types.TargetNetwork

Description

Documentation

data TargetNetwork

Describes a target network associated with a Client VPN endpoint.

See: newTargetNetwork smart constructor.

Constructors

TargetNetwork'

Fields

associationId :: Maybe Text

The ID of the association.

status :: Maybe AssociationStatus

The current state of the target network association.

securityGroups :: Maybe [Text]

The IDs of the security groups applied to the target network association.

targetNetworkId :: Maybe Text

The ID of the subnet specified as the target network.

vpcId :: Maybe Text

The ID of the VPC in which the target network (subnet) is located.

clientVpnEndpointId :: Maybe Text

The ID of the Client VPN endpoint with which the target network is associated.

newTargetNetwork :: TargetNetwork

Create a value of TargetNetwork with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:associationId:TargetNetwork', targetNetwork_associationId - The ID of the association.

$sel:status:TargetNetwork', targetNetwork_status - The current state of the target network association.

$sel:securityGroups:TargetNetwork', targetNetwork_securityGroups - The IDs of the security groups applied to the target network association.

$sel:targetNetworkId:TargetNetwork', targetNetwork_targetNetworkId - The ID of the subnet specified as the target network.

$sel:vpcId:TargetNetwork', targetNetwork_vpcId - The ID of the VPC in which the target network (subnet) is located.

$sel:clientVpnEndpointId:TargetNetwork', targetNetwork_clientVpnEndpointId - The ID of the Client VPN endpoint with which the target network is associated.

targetNetwork_associationId :: Lens' TargetNetwork (Maybe Text)

The ID of the association.

targetNetwork_status :: Lens' TargetNetwork (Maybe AssociationStatus)

The current state of the target network association.

targetNetwork_securityGroups :: Lens' TargetNetwork (Maybe [Text])

The IDs of the security groups applied to the target network association.

targetNetwork_targetNetworkId :: Lens' TargetNetwork (Maybe Text)

The ID of the subnet specified as the target network.

targetNetwork_vpcId :: Lens' TargetNetwork (Maybe Text)

The ID of the VPC in which the target network (subnet) is located.

targetNetwork_clientVpnEndpointId :: Lens' TargetNetwork (Maybe Text)

The ID of the Client VPN endpoint with which the target network is associated.