Network.AWS.EC2.Types.TrafficMirrorFilterRule

Description

Documentation

data TrafficMirrorFilterRule

Describes the Traffic Mirror rule.

See: newTrafficMirrorFilterRule smart constructor.

Constructors

TrafficMirrorFilterRule'

Fields

trafficMirrorFilterRuleId :: Maybe Text

The ID of the Traffic Mirror rule.

sourcePortRange :: Maybe TrafficMirrorPortRange

The source port range assigned to the Traffic Mirror rule.

ruleAction :: Maybe TrafficMirrorRuleAction

The action assigned to the Traffic Mirror rule.

sourceCidrBlock :: Maybe Text

The source CIDR block assigned to the Traffic Mirror rule.

trafficDirection :: Maybe TrafficDirection

The traffic direction assigned to the Traffic Mirror rule.

trafficMirrorFilterId :: Maybe Text

The ID of the Traffic Mirror filter that the rule is associated with.

destinationCidrBlock :: Maybe Text

The destination CIDR block assigned to the Traffic Mirror rule.

protocol :: Maybe Int

The protocol assigned to the Traffic Mirror rule.

description :: Maybe Text

The description of the Traffic Mirror rule.

ruleNumber :: Maybe Int

The rule number of the Traffic Mirror rule.

destinationPortRange :: Maybe TrafficMirrorPortRange

The destination port range assigned to the Traffic Mirror rule.

newTrafficMirrorFilterRule :: TrafficMirrorFilterRule

Create a value of TrafficMirrorFilterRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:trafficMirrorFilterRuleId:TrafficMirrorFilterRule', trafficMirrorFilterRule_trafficMirrorFilterRuleId - The ID of the Traffic Mirror rule.

$sel:sourcePortRange:TrafficMirrorFilterRule', trafficMirrorFilterRule_sourcePortRange - The source port range assigned to the Traffic Mirror rule.

$sel:ruleAction:TrafficMirrorFilterRule', trafficMirrorFilterRule_ruleAction - The action assigned to the Traffic Mirror rule.

$sel:sourceCidrBlock:TrafficMirrorFilterRule', trafficMirrorFilterRule_sourceCidrBlock - The source CIDR block assigned to the Traffic Mirror rule.

$sel:trafficDirection:TrafficMirrorFilterRule', trafficMirrorFilterRule_trafficDirection - The traffic direction assigned to the Traffic Mirror rule.

$sel:trafficMirrorFilterId:TrafficMirrorFilterRule', trafficMirrorFilterRule_trafficMirrorFilterId - The ID of the Traffic Mirror filter that the rule is associated with.

$sel:destinationCidrBlock:TrafficMirrorFilterRule', trafficMirrorFilterRule_destinationCidrBlock - The destination CIDR block assigned to the Traffic Mirror rule.

$sel:protocol:TrafficMirrorFilterRule', trafficMirrorFilterRule_protocol - The protocol assigned to the Traffic Mirror rule.

$sel:description:TrafficMirrorFilterRule', trafficMirrorFilterRule_description - The description of the Traffic Mirror rule.

$sel:ruleNumber:TrafficMirrorFilterRule', trafficMirrorFilterRule_ruleNumber - The rule number of the Traffic Mirror rule.

$sel:destinationPortRange:TrafficMirrorFilterRule', trafficMirrorFilterRule_destinationPortRange - The destination port range assigned to the Traffic Mirror rule.

trafficMirrorFilterRule_trafficMirrorFilterRuleId :: Lens' TrafficMirrorFilterRule (Maybe Text)

The ID of the Traffic Mirror rule.

trafficMirrorFilterRule_sourcePortRange :: Lens' TrafficMirrorFilterRule (Maybe TrafficMirrorPortRange)

The source port range assigned to the Traffic Mirror rule.

trafficMirrorFilterRule_ruleAction :: Lens' TrafficMirrorFilterRule (Maybe TrafficMirrorRuleAction)

The action assigned to the Traffic Mirror rule.

trafficMirrorFilterRule_sourceCidrBlock :: Lens' TrafficMirrorFilterRule (Maybe Text)

The source CIDR block assigned to the Traffic Mirror rule.

trafficMirrorFilterRule_trafficDirection :: Lens' TrafficMirrorFilterRule (Maybe TrafficDirection)

The traffic direction assigned to the Traffic Mirror rule.

trafficMirrorFilterRule_trafficMirrorFilterId :: Lens' TrafficMirrorFilterRule (Maybe Text)

The ID of the Traffic Mirror filter that the rule is associated with.

trafficMirrorFilterRule_destinationCidrBlock :: Lens' TrafficMirrorFilterRule (Maybe Text)

The destination CIDR block assigned to the Traffic Mirror rule.

trafficMirrorFilterRule_protocol :: Lens' TrafficMirrorFilterRule (Maybe Int)

The protocol assigned to the Traffic Mirror rule.

trafficMirrorFilterRule_description :: Lens' TrafficMirrorFilterRule (Maybe Text)

The description of the Traffic Mirror rule.

trafficMirrorFilterRule_ruleNumber :: Lens' TrafficMirrorFilterRule (Maybe Int)

The rule number of the Traffic Mirror rule.

trafficMirrorFilterRule_destinationPortRange :: Lens' TrafficMirrorFilterRule (Maybe TrafficMirrorPortRange)

The destination port range assigned to the Traffic Mirror rule.