Network.AWS.EC2.Types.TrafficMirrorSession

Description

Documentation

data TrafficMirrorSession

Describes a Traffic Mirror session.

See: newTrafficMirrorSession smart constructor.

Constructors

TrafficMirrorSession'

Fields

trafficMirrorTargetId :: Maybe Text

The ID of the Traffic Mirror target.

networkInterfaceId :: Maybe Text

The ID of the Traffic Mirror session's network interface.

trafficMirrorFilterId :: Maybe Text

The ID of the Traffic Mirror filter.

packetLength :: Maybe Int

The number of bytes in each packet to mirror. These are the bytes after the VXLAN header. To mirror a subset, set this to the length (in bytes) to mirror. For example, if you set this value to 100, then the first 100 bytes that meet the filter criteria are copied to the target. Do not specify this parameter when you want to mirror the entire packet

ownerId :: Maybe Text

The ID of the account that owns the Traffic Mirror session.

trafficMirrorSessionId :: Maybe Text

The ID for the Traffic Mirror session.

virtualNetworkId :: Maybe Int

The virtual network ID associated with the Traffic Mirror session.

sessionNumber :: Maybe Int

The session number determines the order in which sessions are evaluated when an interface is used by multiple sessions. The first session with a matching filter is the one that mirrors the packets.

Valid values are 1-32766.

description :: Maybe Text

The description of the Traffic Mirror session.

tags :: Maybe [Tag]

The tags assigned to the Traffic Mirror session.

newTrafficMirrorSession :: TrafficMirrorSession

Create a value of TrafficMirrorSession with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:trafficMirrorTargetId:TrafficMirrorSession', trafficMirrorSession_trafficMirrorTargetId - The ID of the Traffic Mirror target.

$sel:networkInterfaceId:TrafficMirrorSession', trafficMirrorSession_networkInterfaceId - The ID of the Traffic Mirror session's network interface.

$sel:trafficMirrorFilterId:TrafficMirrorSession', trafficMirrorSession_trafficMirrorFilterId - The ID of the Traffic Mirror filter.

$sel:packetLength:TrafficMirrorSession', trafficMirrorSession_packetLength - The number of bytes in each packet to mirror. These are the bytes after the VXLAN header. To mirror a subset, set this to the length (in bytes) to mirror. For example, if you set this value to 100, then the first 100 bytes that meet the filter criteria are copied to the target. Do not specify this parameter when you want to mirror the entire packet

$sel:ownerId:TrafficMirrorSession', trafficMirrorSession_ownerId - The ID of the account that owns the Traffic Mirror session.

$sel:trafficMirrorSessionId:TrafficMirrorSession', trafficMirrorSession_trafficMirrorSessionId - The ID for the Traffic Mirror session.

$sel:virtualNetworkId:TrafficMirrorSession', trafficMirrorSession_virtualNetworkId - The virtual network ID associated with the Traffic Mirror session.

$sel:sessionNumber:TrafficMirrorSession', trafficMirrorSession_sessionNumber - The session number determines the order in which sessions are evaluated when an interface is used by multiple sessions. The first session with a matching filter is the one that mirrors the packets.

Valid values are 1-32766.

$sel:description:TrafficMirrorSession', trafficMirrorSession_description - The description of the Traffic Mirror session.

$sel:tags:TrafficMirrorSession', trafficMirrorSession_tags - The tags assigned to the Traffic Mirror session.

trafficMirrorSession_trafficMirrorTargetId :: Lens' TrafficMirrorSession (Maybe Text)

The ID of the Traffic Mirror target.

trafficMirrorSession_networkInterfaceId :: Lens' TrafficMirrorSession (Maybe Text)

The ID of the Traffic Mirror session's network interface.

trafficMirrorSession_trafficMirrorFilterId :: Lens' TrafficMirrorSession (Maybe Text)

The ID of the Traffic Mirror filter.

trafficMirrorSession_packetLength :: Lens' TrafficMirrorSession (Maybe Int)

The number of bytes in each packet to mirror. These are the bytes after the VXLAN header. To mirror a subset, set this to the length (in bytes) to mirror. For example, if you set this value to 100, then the first 100 bytes that meet the filter criteria are copied to the target. Do not specify this parameter when you want to mirror the entire packet

trafficMirrorSession_ownerId :: Lens' TrafficMirrorSession (Maybe Text)

The ID of the account that owns the Traffic Mirror session.

trafficMirrorSession_trafficMirrorSessionId :: Lens' TrafficMirrorSession (Maybe Text)

The ID for the Traffic Mirror session.

trafficMirrorSession_virtualNetworkId :: Lens' TrafficMirrorSession (Maybe Int)

The virtual network ID associated with the Traffic Mirror session.

trafficMirrorSession_sessionNumber :: Lens' TrafficMirrorSession (Maybe Int)

The session number determines the order in which sessions are evaluated when an interface is used by multiple sessions. The first session with a matching filter is the one that mirrors the packets.

Valid values are 1-32766.

trafficMirrorSession_description :: Lens' TrafficMirrorSession (Maybe Text)

The description of the Traffic Mirror session.

trafficMirrorSession_tags :: Lens' TrafficMirrorSession (Maybe [Tag])

The tags assigned to the Traffic Mirror session.