Network.AWS.EC2.Types.TransitGateway

Description

Documentation

data TransitGateway

Describes a transit gateway.

See: newTransitGateway smart constructor.

Constructors

TransitGateway'

Fields

creationTime :: Maybe ISO8601

The creation time.

state :: Maybe TransitGatewayState

The state of the transit gateway.

ownerId :: Maybe Text

The ID of the Amazon Web Services account that owns the transit gateway.

transitGatewayArn :: Maybe Text

The Amazon Resource Name (ARN) of the transit gateway.

transitGatewayId :: Maybe Text

The ID of the transit gateway.

options :: Maybe TransitGatewayOptions

The transit gateway options.

description :: Maybe Text

The description of the transit gateway.

tags :: Maybe [Tag]

The tags for the transit gateway.

newTransitGateway :: TransitGateway

Create a value of TransitGateway with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:TransitGateway', transitGateway_creationTime - The creation time.

$sel:state:TransitGateway', transitGateway_state - The state of the transit gateway.

$sel:ownerId:TransitGateway', transitGateway_ownerId - The ID of the Amazon Web Services account that owns the transit gateway.

$sel:transitGatewayArn:TransitGateway', transitGateway_transitGatewayArn - The Amazon Resource Name (ARN) of the transit gateway.

$sel:transitGatewayId:TransitGateway', transitGateway_transitGatewayId - The ID of the transit gateway.

$sel:options:TransitGateway', transitGateway_options - The transit gateway options.

$sel:description:TransitGateway', transitGateway_description - The description of the transit gateway.

$sel:tags:TransitGateway', transitGateway_tags - The tags for the transit gateway.

transitGateway_creationTime :: Lens' TransitGateway (Maybe UTCTime)

The creation time.

transitGateway_state :: Lens' TransitGateway (Maybe TransitGatewayState)

The state of the transit gateway.

transitGateway_ownerId :: Lens' TransitGateway (Maybe Text)

The ID of the Amazon Web Services account that owns the transit gateway.

transitGateway_transitGatewayArn :: Lens' TransitGateway (Maybe Text)

The Amazon Resource Name (ARN) of the transit gateway.

transitGateway_transitGatewayId :: Lens' TransitGateway (Maybe Text)

The ID of the transit gateway.

transitGateway_options :: Lens' TransitGateway (Maybe TransitGatewayOptions)

The transit gateway options.

transitGateway_description :: Lens' TransitGateway (Maybe Text)

The description of the transit gateway.

transitGateway_tags :: Lens' TransitGateway (Maybe [Tag])

The tags for the transit gateway.