Network.AWS.EC2.Types.TransitGatewayAttachment

Description

Documentation

data TransitGatewayAttachment

Describes an attachment between a resource and a transit gateway.

See: newTransitGatewayAttachment smart constructor.

Constructors

TransitGatewayAttachment'

Fields

creationTime :: Maybe ISO8601

The creation time.

state :: Maybe TransitGatewayAttachmentState

The attachment state. Note that the initiating state has been deprecated.

resourceId :: Maybe Text

The ID of the resource.

resourceType :: Maybe TransitGatewayAttachmentResourceType

The resource type. Note that the tgw-peering resource type has been deprecated.

transitGatewayOwnerId :: Maybe Text

The ID of the Amazon Web Services account that owns the transit gateway.

transitGatewayId :: Maybe Text

The ID of the transit gateway.

transitGatewayAttachmentId :: Maybe Text

The ID of the attachment.

resourceOwnerId :: Maybe Text

The ID of the Amazon Web Services account that owns the resource.

tags :: Maybe [Tag]

The tags for the attachment.

association :: Maybe TransitGatewayAttachmentAssociation

The association.

newTransitGatewayAttachment :: TransitGatewayAttachment

Create a value of TransitGatewayAttachment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:TransitGatewayAttachment', transitGatewayAttachment_creationTime - The creation time.

$sel:state:TransitGatewayAttachment', transitGatewayAttachment_state - The attachment state. Note that the initiating state has been deprecated.

$sel:resourceId:TransitGatewayAttachment', transitGatewayAttachment_resourceId - The ID of the resource.

$sel:resourceType:TransitGatewayAttachment', transitGatewayAttachment_resourceType - The resource type. Note that the tgw-peering resource type has been deprecated.

$sel:transitGatewayOwnerId:TransitGatewayAttachment', transitGatewayAttachment_transitGatewayOwnerId - The ID of the Amazon Web Services account that owns the transit gateway.

$sel:transitGatewayId:TransitGatewayAttachment', transitGatewayAttachment_transitGatewayId - The ID of the transit gateway.

$sel:transitGatewayAttachmentId:TransitGatewayAttachment', transitGatewayAttachment_transitGatewayAttachmentId - The ID of the attachment.

$sel:resourceOwnerId:TransitGatewayAttachment', transitGatewayAttachment_resourceOwnerId - The ID of the Amazon Web Services account that owns the resource.

$sel:tags:TransitGatewayAttachment', transitGatewayAttachment_tags - The tags for the attachment.

$sel:association:TransitGatewayAttachment', transitGatewayAttachment_association - The association.

transitGatewayAttachment_creationTime :: Lens' TransitGatewayAttachment (Maybe UTCTime)

The creation time.

transitGatewayAttachment_state :: Lens' TransitGatewayAttachment (Maybe TransitGatewayAttachmentState)

The attachment state. Note that the initiating state has been deprecated.

transitGatewayAttachment_resourceId :: Lens' TransitGatewayAttachment (Maybe Text)

The ID of the resource.

transitGatewayAttachment_resourceType :: Lens' TransitGatewayAttachment (Maybe TransitGatewayAttachmentResourceType)

The resource type. Note that the tgw-peering resource type has been deprecated.

transitGatewayAttachment_transitGatewayOwnerId :: Lens' TransitGatewayAttachment (Maybe Text)

The ID of the Amazon Web Services account that owns the transit gateway.

transitGatewayAttachment_transitGatewayId :: Lens' TransitGatewayAttachment (Maybe Text)

The ID of the transit gateway.

transitGatewayAttachment_transitGatewayAttachmentId :: Lens' TransitGatewayAttachment (Maybe Text)

The ID of the attachment.

transitGatewayAttachment_resourceOwnerId :: Lens' TransitGatewayAttachment (Maybe Text)

The ID of the Amazon Web Services account that owns the resource.

transitGatewayAttachment_tags :: Lens' TransitGatewayAttachment (Maybe [Tag])

The tags for the attachment.

transitGatewayAttachment_association :: Lens' TransitGatewayAttachment (Maybe TransitGatewayAttachmentAssociation)

The association.