Network.AWS.EC2.Types.TransitGatewayConnect

Description

Documentation

data TransitGatewayConnect

Describes a transit gateway Connect attachment.

See: newTransitGatewayConnect smart constructor.

Constructors

TransitGatewayConnect'

Fields

creationTime :: Maybe ISO8601

The creation time.

options :: Maybe TransitGatewayConnectOptions

The Connect attachment options.

state :: Maybe TransitGatewayAttachmentState

The state of the attachment.

tags :: Maybe [Tag]

The tags for the attachment.

transitGatewayAttachmentId :: Maybe Text

The ID of the Connect attachment.

transportTransitGatewayAttachmentId :: Maybe Text

The ID of the attachment from which the Connect attachment was created.

transitGatewayId :: Maybe Text

The ID of the transit gateway.

newTransitGatewayConnect :: TransitGatewayConnect

Create a value of TransitGatewayConnect with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:TransitGatewayConnect', transitGatewayConnect_creationTime - The creation time.

$sel:options:TransitGatewayConnect', transitGatewayConnect_options - The Connect attachment options.

$sel:state:TransitGatewayConnect', transitGatewayConnect_state - The state of the attachment.

$sel:tags:TransitGatewayConnect', transitGatewayConnect_tags - The tags for the attachment.

$sel:transitGatewayAttachmentId:TransitGatewayConnect', transitGatewayConnect_transitGatewayAttachmentId - The ID of the Connect attachment.

$sel:transportTransitGatewayAttachmentId:TransitGatewayConnect', transitGatewayConnect_transportTransitGatewayAttachmentId - The ID of the attachment from which the Connect attachment was created.

$sel:transitGatewayId:TransitGatewayConnect', transitGatewayConnect_transitGatewayId - The ID of the transit gateway.

transitGatewayConnect_creationTime :: Lens' TransitGatewayConnect (Maybe UTCTime)

The creation time.

transitGatewayConnect_options :: Lens' TransitGatewayConnect (Maybe TransitGatewayConnectOptions)

The Connect attachment options.

transitGatewayConnect_state :: Lens' TransitGatewayConnect (Maybe TransitGatewayAttachmentState)

The state of the attachment.

transitGatewayConnect_tags :: Lens' TransitGatewayConnect (Maybe [Tag])

The tags for the attachment.

transitGatewayConnect_transitGatewayAttachmentId :: Lens' TransitGatewayConnect (Maybe Text)

The ID of the Connect attachment.

transitGatewayConnect_transportTransitGatewayAttachmentId :: Lens' TransitGatewayConnect (Maybe Text)

The ID of the attachment from which the Connect attachment was created.

transitGatewayConnect_transitGatewayId :: Lens' TransitGatewayConnect (Maybe Text)

The ID of the transit gateway.