Network.AWS.EC2.Types.TransitGatewayConnectPeerConfiguration

Description

Documentation

data TransitGatewayConnectPeerConfiguration

Describes the Connect peer details.

See: newTransitGatewayConnectPeerConfiguration smart constructor.

Constructors

TransitGatewayConnectPeerConfiguration'

Fields

protocol :: Maybe ProtocolValue

The tunnel protocol.

transitGatewayAddress :: Maybe Text

The Connect peer IP address on the transit gateway side of the tunnel.

peerAddress :: Maybe Text

The Connect peer IP address on the appliance side of the tunnel.

insideCidrBlocks :: Maybe [Text]

The range of interior BGP peer IP addresses.

bgpConfigurations :: Maybe [TransitGatewayAttachmentBgpConfiguration]

The BGP configuration details.

newTransitGatewayConnectPeerConfiguration :: TransitGatewayConnectPeerConfiguration

Create a value of TransitGatewayConnectPeerConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:protocol:TransitGatewayConnectPeerConfiguration', transitGatewayConnectPeerConfiguration_protocol - The tunnel protocol.

$sel:transitGatewayAddress:TransitGatewayConnectPeerConfiguration', transitGatewayConnectPeerConfiguration_transitGatewayAddress - The Connect peer IP address on the transit gateway side of the tunnel.

$sel:peerAddress:TransitGatewayConnectPeerConfiguration', transitGatewayConnectPeerConfiguration_peerAddress - The Connect peer IP address on the appliance side of the tunnel.

$sel:insideCidrBlocks:TransitGatewayConnectPeerConfiguration', transitGatewayConnectPeerConfiguration_insideCidrBlocks - The range of interior BGP peer IP addresses.

$sel:bgpConfigurations:TransitGatewayConnectPeerConfiguration', transitGatewayConnectPeerConfiguration_bgpConfigurations - The BGP configuration details.

transitGatewayConnectPeerConfiguration_protocol :: Lens' TransitGatewayConnectPeerConfiguration (Maybe ProtocolValue)

The tunnel protocol.

transitGatewayConnectPeerConfiguration_transitGatewayAddress :: Lens' TransitGatewayConnectPeerConfiguration (Maybe Text)

The Connect peer IP address on the transit gateway side of the tunnel.

transitGatewayConnectPeerConfiguration_peerAddress :: Lens' TransitGatewayConnectPeerConfiguration (Maybe Text)

The Connect peer IP address on the appliance side of the tunnel.

transitGatewayConnectPeerConfiguration_insideCidrBlocks :: Lens' TransitGatewayConnectPeerConfiguration (Maybe [Text])

The range of interior BGP peer IP addresses.

transitGatewayConnectPeerConfiguration_bgpConfigurations :: Lens' TransitGatewayConnectPeerConfiguration (Maybe [TransitGatewayAttachmentBgpConfiguration])

The BGP configuration details.