Network.AWS.EC2.Types.TransitGatewayMulticastDomainAssociations

Description

Documentation

data TransitGatewayMulticastDomainAssociations

Describes the multicast domain associations.

See: newTransitGatewayMulticastDomainAssociations smart constructor.

Constructors

TransitGatewayMulticastDomainAssociations'

Fields

resourceId :: Maybe Text

The ID of the resource.

resourceType :: Maybe TransitGatewayAttachmentResourceType

The type of resource, for example a VPC attachment.

subnets :: Maybe [SubnetAssociation]

The subnets associated with the multicast domain.

transitGatewayMulticastDomainId :: Maybe Text

The ID of the transit gateway multicast domain.

transitGatewayAttachmentId :: Maybe Text

The ID of the transit gateway attachment.

resourceOwnerId :: Maybe Text

The ID of the Amazon Web Services account that owns the resource.

newTransitGatewayMulticastDomainAssociations :: TransitGatewayMulticastDomainAssociations

Create a value of TransitGatewayMulticastDomainAssociations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceId:TransitGatewayMulticastDomainAssociations', transitGatewayMulticastDomainAssociations_resourceId - The ID of the resource.

$sel:resourceType:TransitGatewayMulticastDomainAssociations', transitGatewayMulticastDomainAssociations_resourceType - The type of resource, for example a VPC attachment.

$sel:subnets:TransitGatewayMulticastDomainAssociations', transitGatewayMulticastDomainAssociations_subnets - The subnets associated with the multicast domain.

$sel:transitGatewayMulticastDomainId:TransitGatewayMulticastDomainAssociations', transitGatewayMulticastDomainAssociations_transitGatewayMulticastDomainId - The ID of the transit gateway multicast domain.

$sel:transitGatewayAttachmentId:TransitGatewayMulticastDomainAssociations', transitGatewayMulticastDomainAssociations_transitGatewayAttachmentId - The ID of the transit gateway attachment.

$sel:resourceOwnerId:TransitGatewayMulticastDomainAssociations', transitGatewayMulticastDomainAssociations_resourceOwnerId - The ID of the Amazon Web Services account that owns the resource.

transitGatewayMulticastDomainAssociations_resourceId :: Lens' TransitGatewayMulticastDomainAssociations (Maybe Text)

The ID of the resource.

transitGatewayMulticastDomainAssociations_resourceType :: Lens' TransitGatewayMulticastDomainAssociations (Maybe TransitGatewayAttachmentResourceType)

The type of resource, for example a VPC attachment.

transitGatewayMulticastDomainAssociations_subnets :: Lens' TransitGatewayMulticastDomainAssociations (Maybe [SubnetAssociation])

The subnets associated with the multicast domain.

transitGatewayMulticastDomainAssociations_transitGatewayMulticastDomainId :: Lens' TransitGatewayMulticastDomainAssociations (Maybe Text)

The ID of the transit gateway multicast domain.

transitGatewayMulticastDomainAssociations_transitGatewayAttachmentId :: Lens' TransitGatewayMulticastDomainAssociations (Maybe Text)

The ID of the transit gateway attachment.

transitGatewayMulticastDomainAssociations_resourceOwnerId :: Lens' TransitGatewayMulticastDomainAssociations (Maybe Text)

The ID of the Amazon Web Services account that owns the resource.