Network.AWS.EC2.Types.TransitGatewayMulticastGroup

Description

Documentation

data TransitGatewayMulticastGroup

Describes the transit gateway multicast group resources.

See: newTransitGatewayMulticastGroup smart constructor.

Constructors

TransitGatewayMulticastGroup'

Fields

resourceId :: Maybe Text

The ID of the resource.

resourceType :: Maybe TransitGatewayAttachmentResourceType

The type of resource, for example a VPC attachment.

sourceType :: Maybe MembershipType

The source type.

memberType :: Maybe MembershipType

The member type (for example, static).

networkInterfaceId :: Maybe Text

The ID of the transit gateway attachment.

subnetId :: Maybe Text

The ID of the subnet.

groupMember :: Maybe Bool

Indicates that the resource is a transit gateway multicast group member.

groupSource :: Maybe Bool

Indicates that the resource is a transit gateway multicast group member.

groupIpAddress :: Maybe Text

The IP address assigned to the transit gateway multicast group.

transitGatewayAttachmentId :: Maybe Text

The ID of the transit gateway attachment.

resourceOwnerId :: Maybe Text

The ID of the Amazon Web Services account that owns the transit gateway multicast domain group resource.

newTransitGatewayMulticastGroup :: TransitGatewayMulticastGroup

Create a value of TransitGatewayMulticastGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceId:TransitGatewayMulticastGroup', transitGatewayMulticastGroup_resourceId - The ID of the resource.

$sel:resourceType:TransitGatewayMulticastGroup', transitGatewayMulticastGroup_resourceType - The type of resource, for example a VPC attachment.

$sel:sourceType:TransitGatewayMulticastGroup', transitGatewayMulticastGroup_sourceType - The source type.

$sel:memberType:TransitGatewayMulticastGroup', transitGatewayMulticastGroup_memberType - The member type (for example, static).

$sel:networkInterfaceId:TransitGatewayMulticastGroup', transitGatewayMulticastGroup_networkInterfaceId - The ID of the transit gateway attachment.

$sel:subnetId:TransitGatewayMulticastGroup', transitGatewayMulticastGroup_subnetId - The ID of the subnet.

$sel:groupMember:TransitGatewayMulticastGroup', transitGatewayMulticastGroup_groupMember - Indicates that the resource is a transit gateway multicast group member.

$sel:groupSource:TransitGatewayMulticastGroup', transitGatewayMulticastGroup_groupSource - Indicates that the resource is a transit gateway multicast group member.

$sel:groupIpAddress:TransitGatewayMulticastGroup', transitGatewayMulticastGroup_groupIpAddress - The IP address assigned to the transit gateway multicast group.

$sel:transitGatewayAttachmentId:TransitGatewayMulticastGroup', transitGatewayMulticastGroup_transitGatewayAttachmentId - The ID of the transit gateway attachment.

$sel:resourceOwnerId:TransitGatewayMulticastGroup', transitGatewayMulticastGroup_resourceOwnerId - The ID of the Amazon Web Services account that owns the transit gateway multicast domain group resource.

transitGatewayMulticastGroup_resourceId :: Lens' TransitGatewayMulticastGroup (Maybe Text)

The ID of the resource.

transitGatewayMulticastGroup_resourceType :: Lens' TransitGatewayMulticastGroup (Maybe TransitGatewayAttachmentResourceType)

The type of resource, for example a VPC attachment.

transitGatewayMulticastGroup_sourceType :: Lens' TransitGatewayMulticastGroup (Maybe MembershipType)

The source type.

transitGatewayMulticastGroup_memberType :: Lens' TransitGatewayMulticastGroup (Maybe MembershipType)

The member type (for example, static).

transitGatewayMulticastGroup_networkInterfaceId :: Lens' TransitGatewayMulticastGroup (Maybe Text)

The ID of the transit gateway attachment.

transitGatewayMulticastGroup_subnetId :: Lens' TransitGatewayMulticastGroup (Maybe Text)

The ID of the subnet.

transitGatewayMulticastGroup_groupMember :: Lens' TransitGatewayMulticastGroup (Maybe Bool)

Indicates that the resource is a transit gateway multicast group member.

transitGatewayMulticastGroup_groupSource :: Lens' TransitGatewayMulticastGroup (Maybe Bool)

Indicates that the resource is a transit gateway multicast group member.

transitGatewayMulticastGroup_groupIpAddress :: Lens' TransitGatewayMulticastGroup (Maybe Text)

The IP address assigned to the transit gateway multicast group.

transitGatewayMulticastGroup_transitGatewayAttachmentId :: Lens' TransitGatewayMulticastGroup (Maybe Text)

The ID of the transit gateway attachment.

transitGatewayMulticastGroup_resourceOwnerId :: Lens' TransitGatewayMulticastGroup (Maybe Text)

The ID of the Amazon Web Services account that owns the transit gateway multicast domain group resource.