Network.AWS.EC2.Types.TransitGatewayOptions

Description

Documentation

data TransitGatewayOptions

Describes the options for a transit gateway.

See: newTransitGatewayOptions smart constructor.

Constructors

TransitGatewayOptions'

Fields

propagationDefaultRouteTableId :: Maybe Text

The ID of the default propagation route table.

vpnEcmpSupport :: Maybe VpnEcmpSupportValue

Indicates whether Equal Cost Multipath Protocol support is enabled.

dnsSupport :: Maybe DnsSupportValue

Indicates whether DNS support is enabled.

associationDefaultRouteTableId :: Maybe Text

The ID of the default association route table.

autoAcceptSharedAttachments :: Maybe AutoAcceptSharedAttachmentsValue

Indicates whether attachment requests are automatically accepted.

transitGatewayCidrBlocks :: Maybe [Text]

The transit gateway CIDR blocks.

amazonSideAsn :: Maybe Integer

A private Autonomous System Number (ASN) for the Amazon side of a BGP session. The range is 64512 to 65534 for 16-bit ASNs and 4200000000 to 4294967294 for 32-bit ASNs.

multicastSupport :: Maybe MulticastSupportValue

Indicates whether multicast is enabled on the transit gateway

defaultRouteTableAssociation :: Maybe DefaultRouteTableAssociationValue

Indicates whether resource attachments are automatically associated with the default association route table.

defaultRouteTablePropagation :: Maybe DefaultRouteTablePropagationValue

Indicates whether resource attachments automatically propagate routes to the default propagation route table.

newTransitGatewayOptions :: TransitGatewayOptions

Create a value of TransitGatewayOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:propagationDefaultRouteTableId:TransitGatewayOptions', transitGatewayOptions_propagationDefaultRouteTableId - The ID of the default propagation route table.

$sel:vpnEcmpSupport:TransitGatewayOptions', transitGatewayOptions_vpnEcmpSupport - Indicates whether Equal Cost Multipath Protocol support is enabled.

$sel:dnsSupport:TransitGatewayOptions', transitGatewayOptions_dnsSupport - Indicates whether DNS support is enabled.

$sel:associationDefaultRouteTableId:TransitGatewayOptions', transitGatewayOptions_associationDefaultRouteTableId - The ID of the default association route table.

$sel:autoAcceptSharedAttachments:TransitGatewayOptions', transitGatewayOptions_autoAcceptSharedAttachments - Indicates whether attachment requests are automatically accepted.

$sel:transitGatewayCidrBlocks:TransitGatewayOptions', transitGatewayOptions_transitGatewayCidrBlocks - The transit gateway CIDR blocks.

$sel:amazonSideAsn:TransitGatewayOptions', transitGatewayOptions_amazonSideAsn - A private Autonomous System Number (ASN) for the Amazon side of a BGP session. The range is 64512 to 65534 for 16-bit ASNs and 4200000000 to 4294967294 for 32-bit ASNs.

$sel:multicastSupport:TransitGatewayOptions', transitGatewayOptions_multicastSupport - Indicates whether multicast is enabled on the transit gateway

$sel:defaultRouteTableAssociation:TransitGatewayOptions', transitGatewayOptions_defaultRouteTableAssociation - Indicates whether resource attachments are automatically associated with the default association route table.

$sel:defaultRouteTablePropagation:TransitGatewayOptions', transitGatewayOptions_defaultRouteTablePropagation - Indicates whether resource attachments automatically propagate routes to the default propagation route table.

transitGatewayOptions_propagationDefaultRouteTableId :: Lens' TransitGatewayOptions (Maybe Text)

The ID of the default propagation route table.

transitGatewayOptions_vpnEcmpSupport :: Lens' TransitGatewayOptions (Maybe VpnEcmpSupportValue)

Indicates whether Equal Cost Multipath Protocol support is enabled.

transitGatewayOptions_dnsSupport :: Lens' TransitGatewayOptions (Maybe DnsSupportValue)

Indicates whether DNS support is enabled.

transitGatewayOptions_associationDefaultRouteTableId :: Lens' TransitGatewayOptions (Maybe Text)

The ID of the default association route table.

transitGatewayOptions_autoAcceptSharedAttachments :: Lens' TransitGatewayOptions (Maybe AutoAcceptSharedAttachmentsValue)

Indicates whether attachment requests are automatically accepted.

transitGatewayOptions_transitGatewayCidrBlocks :: Lens' TransitGatewayOptions (Maybe [Text])

The transit gateway CIDR blocks.

transitGatewayOptions_amazonSideAsn :: Lens' TransitGatewayOptions (Maybe Integer)

A private Autonomous System Number (ASN) for the Amazon side of a BGP session. The range is 64512 to 65534 for 16-bit ASNs and 4200000000 to 4294967294 for 32-bit ASNs.

transitGatewayOptions_multicastSupport :: Lens' TransitGatewayOptions (Maybe MulticastSupportValue)

Indicates whether multicast is enabled on the transit gateway

transitGatewayOptions_defaultRouteTableAssociation :: Lens' TransitGatewayOptions (Maybe DefaultRouteTableAssociationValue)

Indicates whether resource attachments are automatically associated with the default association route table.

transitGatewayOptions_defaultRouteTablePropagation :: Lens' TransitGatewayOptions (Maybe DefaultRouteTablePropagationValue)

Indicates whether resource attachments automatically propagate routes to the default propagation route table.