Network.AWS.EC2.Types.TransitGatewayPeeringAttachment

Description

Documentation

data TransitGatewayPeeringAttachment

Describes the transit gateway peering attachment.

See: newTransitGatewayPeeringAttachment smart constructor.

Constructors

TransitGatewayPeeringAttachment'

Fields

creationTime :: Maybe ISO8601

The time the transit gateway peering attachment was created.

requesterTgwInfo :: Maybe PeeringTgwInfo

Information about the requester transit gateway.

status :: Maybe PeeringAttachmentStatus

The status of the transit gateway peering attachment.

accepterTgwInfo :: Maybe PeeringTgwInfo

Information about the accepter transit gateway.

state :: Maybe TransitGatewayAttachmentState

The state of the transit gateway peering attachment. Note that the initiating state has been deprecated.

tags :: Maybe [Tag]

The tags for the transit gateway peering attachment.

transitGatewayAttachmentId :: Maybe Text

The ID of the transit gateway peering attachment.

newTransitGatewayPeeringAttachment :: TransitGatewayPeeringAttachment

Create a value of TransitGatewayPeeringAttachment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:TransitGatewayPeeringAttachment', transitGatewayPeeringAttachment_creationTime - The time the transit gateway peering attachment was created.

$sel:requesterTgwInfo:TransitGatewayPeeringAttachment', transitGatewayPeeringAttachment_requesterTgwInfo - Information about the requester transit gateway.

$sel:status:TransitGatewayPeeringAttachment', transitGatewayPeeringAttachment_status - The status of the transit gateway peering attachment.

$sel:accepterTgwInfo:TransitGatewayPeeringAttachment', transitGatewayPeeringAttachment_accepterTgwInfo - Information about the accepter transit gateway.

$sel:state:TransitGatewayPeeringAttachment', transitGatewayPeeringAttachment_state - The state of the transit gateway peering attachment. Note that the initiating state has been deprecated.

$sel:tags:TransitGatewayPeeringAttachment', transitGatewayPeeringAttachment_tags - The tags for the transit gateway peering attachment.

$sel:transitGatewayAttachmentId:TransitGatewayPeeringAttachment', transitGatewayPeeringAttachment_transitGatewayAttachmentId - The ID of the transit gateway peering attachment.

transitGatewayPeeringAttachment_creationTime :: Lens' TransitGatewayPeeringAttachment (Maybe UTCTime)

The time the transit gateway peering attachment was created.

transitGatewayPeeringAttachment_requesterTgwInfo :: Lens' TransitGatewayPeeringAttachment (Maybe PeeringTgwInfo)

Information about the requester transit gateway.

transitGatewayPeeringAttachment_status :: Lens' TransitGatewayPeeringAttachment (Maybe PeeringAttachmentStatus)

The status of the transit gateway peering attachment.

transitGatewayPeeringAttachment_accepterTgwInfo :: Lens' TransitGatewayPeeringAttachment (Maybe PeeringTgwInfo)

Information about the accepter transit gateway.

transitGatewayPeeringAttachment_state :: Lens' TransitGatewayPeeringAttachment (Maybe TransitGatewayAttachmentState)

The state of the transit gateway peering attachment. Note that the initiating state has been deprecated.

transitGatewayPeeringAttachment_tags :: Lens' TransitGatewayPeeringAttachment (Maybe [Tag])

The tags for the transit gateway peering attachment.

transitGatewayPeeringAttachment_transitGatewayAttachmentId :: Lens' TransitGatewayPeeringAttachment (Maybe Text)

The ID of the transit gateway peering attachment.