Network.AWS.EC2.Types.TransitGatewayRequestOptions

Description

Documentation

data TransitGatewayRequestOptions

Describes the options for a transit gateway.

See: newTransitGatewayRequestOptions smart constructor.

Constructors

TransitGatewayRequestOptions'

Fields

vpnEcmpSupport :: Maybe VpnEcmpSupportValue

Enable or disable Equal Cost Multipath Protocol support. Enabled by default.

dnsSupport :: Maybe DnsSupportValue

Enable or disable DNS support. Enabled by default.

autoAcceptSharedAttachments :: Maybe AutoAcceptSharedAttachmentsValue

Enable or disable automatic acceptance of attachment requests. Disabled by default.

transitGatewayCidrBlocks :: Maybe [Text]

One or more IPv4 or IPv6 CIDR blocks for the transit gateway. Must be a size /24 CIDR block or larger for IPv4, or a size /64 CIDR block or larger for IPv6.

amazonSideAsn :: Maybe Integer

A private Autonomous System Number (ASN) for the Amazon side of a BGP session. The range is 64512 to 65534 for 16-bit ASNs and 4200000000 to 4294967294 for 32-bit ASNs. The default is 64512.

multicastSupport :: Maybe MulticastSupportValue

Indicates whether multicast is enabled on the transit gateway

defaultRouteTableAssociation :: Maybe DefaultRouteTableAssociationValue

Enable or disable automatic association with the default association route table. Enabled by default.

defaultRouteTablePropagation :: Maybe DefaultRouteTablePropagationValue

Enable or disable automatic propagation of routes to the default propagation route table. Enabled by default.

newTransitGatewayRequestOptions :: TransitGatewayRequestOptions

Create a value of TransitGatewayRequestOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpnEcmpSupport:TransitGatewayRequestOptions', transitGatewayRequestOptions_vpnEcmpSupport - Enable or disable Equal Cost Multipath Protocol support. Enabled by default.

$sel:dnsSupport:TransitGatewayRequestOptions', transitGatewayRequestOptions_dnsSupport - Enable or disable DNS support. Enabled by default.

$sel:autoAcceptSharedAttachments:TransitGatewayRequestOptions', transitGatewayRequestOptions_autoAcceptSharedAttachments - Enable or disable automatic acceptance of attachment requests. Disabled by default.

$sel:transitGatewayCidrBlocks:TransitGatewayRequestOptions', transitGatewayRequestOptions_transitGatewayCidrBlocks - One or more IPv4 or IPv6 CIDR blocks for the transit gateway. Must be a size /24 CIDR block or larger for IPv4, or a size /64 CIDR block or larger for IPv6.

$sel:amazonSideAsn:TransitGatewayRequestOptions', transitGatewayRequestOptions_amazonSideAsn - A private Autonomous System Number (ASN) for the Amazon side of a BGP session. The range is 64512 to 65534 for 16-bit ASNs and 4200000000 to 4294967294 for 32-bit ASNs. The default is 64512.

$sel:multicastSupport:TransitGatewayRequestOptions', transitGatewayRequestOptions_multicastSupport - Indicates whether multicast is enabled on the transit gateway

$sel:defaultRouteTableAssociation:TransitGatewayRequestOptions', transitGatewayRequestOptions_defaultRouteTableAssociation - Enable or disable automatic association with the default association route table. Enabled by default.

$sel:defaultRouteTablePropagation:TransitGatewayRequestOptions', transitGatewayRequestOptions_defaultRouteTablePropagation - Enable or disable automatic propagation of routes to the default propagation route table. Enabled by default.

transitGatewayRequestOptions_vpnEcmpSupport :: Lens' TransitGatewayRequestOptions (Maybe VpnEcmpSupportValue)

Enable or disable Equal Cost Multipath Protocol support. Enabled by default.

transitGatewayRequestOptions_dnsSupport :: Lens' TransitGatewayRequestOptions (Maybe DnsSupportValue)

Enable or disable DNS support. Enabled by default.

transitGatewayRequestOptions_autoAcceptSharedAttachments :: Lens' TransitGatewayRequestOptions (Maybe AutoAcceptSharedAttachmentsValue)

Enable or disable automatic acceptance of attachment requests. Disabled by default.

transitGatewayRequestOptions_transitGatewayCidrBlocks :: Lens' TransitGatewayRequestOptions (Maybe [Text])

One or more IPv4 or IPv6 CIDR blocks for the transit gateway. Must be a size /24 CIDR block or larger for IPv4, or a size /64 CIDR block or larger for IPv6.

transitGatewayRequestOptions_amazonSideAsn :: Lens' TransitGatewayRequestOptions (Maybe Integer)

A private Autonomous System Number (ASN) for the Amazon side of a BGP session. The range is 64512 to 65534 for 16-bit ASNs and 4200000000 to 4294967294 for 32-bit ASNs. The default is 64512.

transitGatewayRequestOptions_multicastSupport :: Lens' TransitGatewayRequestOptions (Maybe MulticastSupportValue)

Indicates whether multicast is enabled on the transit gateway

transitGatewayRequestOptions_defaultRouteTableAssociation :: Lens' TransitGatewayRequestOptions (Maybe DefaultRouteTableAssociationValue)

Enable or disable automatic association with the default association route table. Enabled by default.

transitGatewayRequestOptions_defaultRouteTablePropagation :: Lens' TransitGatewayRequestOptions (Maybe DefaultRouteTablePropagationValue)

Enable or disable automatic propagation of routes to the default propagation route table. Enabled by default.