Network.AWS.EC2.Types.TransitGatewayRouteTable

Description

Documentation

data TransitGatewayRouteTable

Describes a transit gateway route table.

See: newTransitGatewayRouteTable smart constructor.

Constructors

TransitGatewayRouteTable'

Fields

creationTime :: Maybe ISO8601

The creation time.

state :: Maybe TransitGatewayRouteTableState

The state of the transit gateway route table.

defaultPropagationRouteTable :: Maybe Bool

Indicates whether this is the default propagation route table for the transit gateway.

transitGatewayRouteTableId :: Maybe Text

The ID of the transit gateway route table.

transitGatewayId :: Maybe Text

The ID of the transit gateway.

defaultAssociationRouteTable :: Maybe Bool

Indicates whether this is the default association route table for the transit gateway.

tags :: Maybe [Tag]

Any tags assigned to the route table.

newTransitGatewayRouteTable :: TransitGatewayRouteTable

Create a value of TransitGatewayRouteTable with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:TransitGatewayRouteTable', transitGatewayRouteTable_creationTime - The creation time.

$sel:state:TransitGatewayRouteTable', transitGatewayRouteTable_state - The state of the transit gateway route table.

$sel:defaultPropagationRouteTable:TransitGatewayRouteTable', transitGatewayRouteTable_defaultPropagationRouteTable - Indicates whether this is the default propagation route table for the transit gateway.

$sel:transitGatewayRouteTableId:TransitGatewayRouteTable', transitGatewayRouteTable_transitGatewayRouteTableId - The ID of the transit gateway route table.

$sel:transitGatewayId:TransitGatewayRouteTable', transitGatewayRouteTable_transitGatewayId - The ID of the transit gateway.

$sel:defaultAssociationRouteTable:TransitGatewayRouteTable', transitGatewayRouteTable_defaultAssociationRouteTable - Indicates whether this is the default association route table for the transit gateway.

$sel:tags:TransitGatewayRouteTable', transitGatewayRouteTable_tags - Any tags assigned to the route table.

transitGatewayRouteTable_creationTime :: Lens' TransitGatewayRouteTable (Maybe UTCTime)

The creation time.

transitGatewayRouteTable_state :: Lens' TransitGatewayRouteTable (Maybe TransitGatewayRouteTableState)

The state of the transit gateway route table.

transitGatewayRouteTable_defaultPropagationRouteTable :: Lens' TransitGatewayRouteTable (Maybe Bool)

Indicates whether this is the default propagation route table for the transit gateway.

transitGatewayRouteTable_transitGatewayRouteTableId :: Lens' TransitGatewayRouteTable (Maybe Text)

The ID of the transit gateway route table.

transitGatewayRouteTable_transitGatewayId :: Lens' TransitGatewayRouteTable (Maybe Text)

The ID of the transit gateway.

transitGatewayRouteTable_defaultAssociationRouteTable :: Lens' TransitGatewayRouteTable (Maybe Bool)

Indicates whether this is the default association route table for the transit gateway.

transitGatewayRouteTable_tags :: Lens' TransitGatewayRouteTable (Maybe [Tag])

Any tags assigned to the route table.