Network.AWS.EC2.Types.TransitGatewayVpcAttachment

Description

Documentation

data TransitGatewayVpcAttachment

Describes a VPC attachment.

See: newTransitGatewayVpcAttachment smart constructor.

Constructors

TransitGatewayVpcAttachment'

Fields

creationTime :: Maybe ISO8601

The creation time.

options :: Maybe TransitGatewayVpcAttachmentOptions

The VPC attachment options.

subnetIds :: Maybe [Text]

The IDs of the subnets.

state :: Maybe TransitGatewayAttachmentState

The state of the VPC attachment. Note that the initiating state has been deprecated.

tags :: Maybe [Tag]

The tags for the VPC attachment.

vpcOwnerId :: Maybe Text

The ID of the Amazon Web Services account that owns the VPC.

transitGatewayAttachmentId :: Maybe Text

The ID of the attachment.

vpcId :: Maybe Text

The ID of the VPC.

transitGatewayId :: Maybe Text

The ID of the transit gateway.

newTransitGatewayVpcAttachment :: TransitGatewayVpcAttachment

Create a value of TransitGatewayVpcAttachment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:TransitGatewayVpcAttachment', transitGatewayVpcAttachment_creationTime - The creation time.

$sel:options:TransitGatewayVpcAttachment', transitGatewayVpcAttachment_options - The VPC attachment options.

$sel:subnetIds:TransitGatewayVpcAttachment', transitGatewayVpcAttachment_subnetIds - The IDs of the subnets.

$sel:state:TransitGatewayVpcAttachment', transitGatewayVpcAttachment_state - The state of the VPC attachment. Note that the initiating state has been deprecated.

$sel:tags:TransitGatewayVpcAttachment', transitGatewayVpcAttachment_tags - The tags for the VPC attachment.

$sel:vpcOwnerId:TransitGatewayVpcAttachment', transitGatewayVpcAttachment_vpcOwnerId - The ID of the Amazon Web Services account that owns the VPC.

$sel:transitGatewayAttachmentId:TransitGatewayVpcAttachment', transitGatewayVpcAttachment_transitGatewayAttachmentId - The ID of the attachment.

$sel:vpcId:TransitGatewayVpcAttachment', transitGatewayVpcAttachment_vpcId - The ID of the VPC.

$sel:transitGatewayId:TransitGatewayVpcAttachment', transitGatewayVpcAttachment_transitGatewayId - The ID of the transit gateway.

transitGatewayVpcAttachment_creationTime :: Lens' TransitGatewayVpcAttachment (Maybe UTCTime)

The creation time.

transitGatewayVpcAttachment_options :: Lens' TransitGatewayVpcAttachment (Maybe TransitGatewayVpcAttachmentOptions)

The VPC attachment options.

transitGatewayVpcAttachment_subnetIds :: Lens' TransitGatewayVpcAttachment (Maybe [Text])

The IDs of the subnets.

transitGatewayVpcAttachment_state :: Lens' TransitGatewayVpcAttachment (Maybe TransitGatewayAttachmentState)

The state of the VPC attachment. Note that the initiating state has been deprecated.

transitGatewayVpcAttachment_tags :: Lens' TransitGatewayVpcAttachment (Maybe [Tag])

The tags for the VPC attachment.

transitGatewayVpcAttachment_vpcOwnerId :: Lens' TransitGatewayVpcAttachment (Maybe Text)

The ID of the Amazon Web Services account that owns the VPC.

transitGatewayVpcAttachment_transitGatewayAttachmentId :: Lens' TransitGatewayVpcAttachment (Maybe Text)

The ID of the attachment.

transitGatewayVpcAttachment_vpcId :: Lens' TransitGatewayVpcAttachment (Maybe Text)

The ID of the VPC.

transitGatewayVpcAttachment_transitGatewayId :: Lens' TransitGatewayVpcAttachment (Maybe Text)

The ID of the transit gateway.