Network.AWS.EC2.Types.UserIdGroupPair

Description

Documentation

data UserIdGroupPair

Describes a security group and Amazon Web Services account ID pair.

See: newUserIdGroupPair smart constructor.

Constructors

UserIdGroupPair'

Fields

vpcPeeringConnectionId :: Maybe Text

The ID of the VPC peering connection, if applicable.

groupName :: Maybe Text

The name of the security group. In a request, use this parameter for a security group in EC2-Classic or a default VPC only. For a security group in a nondefault VPC, use the security group ID.

For a referenced security group in another VPC, this value is not returned if the referenced security group is deleted.

groupId :: Maybe Text

The ID of the security group.

userId :: Maybe Text

The ID of an Amazon Web Services account.

For a referenced security group in another VPC, the account ID of the referenced security group is returned in the response. If the referenced security group is deleted, this value is not returned.

EC2-Classic

Required when adding or removing rules that reference a security group in another Amazon Web Services account.

peeringStatus :: Maybe Text

The status of a VPC peering connection, if applicable.

description :: Maybe Text

A description for the security group rule that references this user ID group pair.

Constraints: Up to 255 characters in length. Allowed characters are a-z, A-Z, 0-9, spaces, and ._-:/()#,@[]+=;{}!$*

vpcId :: Maybe Text

The ID of the VPC for the referenced security group, if applicable.

newUserIdGroupPair :: UserIdGroupPair

Create a value of UserIdGroupPair with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpcPeeringConnectionId:UserIdGroupPair', userIdGroupPair_vpcPeeringConnectionId - The ID of the VPC peering connection, if applicable.

$sel:groupName:UserIdGroupPair', userIdGroupPair_groupName - The name of the security group. In a request, use this parameter for a security group in EC2-Classic or a default VPC only. For a security group in a nondefault VPC, use the security group ID.

For a referenced security group in another VPC, this value is not returned if the referenced security group is deleted.

$sel:groupId:UserIdGroupPair', userIdGroupPair_groupId - The ID of the security group.

$sel:userId:UserIdGroupPair', userIdGroupPair_userId - The ID of an Amazon Web Services account.

For a referenced security group in another VPC, the account ID of the referenced security group is returned in the response. If the referenced security group is deleted, this value is not returned.

EC2-Classic

Required when adding or removing rules that reference a security group in another Amazon Web Services account.

$sel:peeringStatus:UserIdGroupPair', userIdGroupPair_peeringStatus - The status of a VPC peering connection, if applicable.

$sel:description:UserIdGroupPair', userIdGroupPair_description - A description for the security group rule that references this user ID group pair.

Constraints: Up to 255 characters in length. Allowed characters are a-z, A-Z, 0-9, spaces, and ._-:/()#,@[]+=;{}!$*

$sel:vpcId:UserIdGroupPair', userIdGroupPair_vpcId - The ID of the VPC for the referenced security group, if applicable.

userIdGroupPair_vpcPeeringConnectionId :: Lens' UserIdGroupPair (Maybe Text)

The ID of the VPC peering connection, if applicable.

userIdGroupPair_groupName :: Lens' UserIdGroupPair (Maybe Text)

The name of the security group. In a request, use this parameter for a security group in EC2-Classic or a default VPC only. For a security group in a nondefault VPC, use the security group ID.

For a referenced security group in another VPC, this value is not returned if the referenced security group is deleted.

userIdGroupPair_groupId :: Lens' UserIdGroupPair (Maybe Text)

The ID of the security group.

userIdGroupPair_userId :: Lens' UserIdGroupPair (Maybe Text)

The ID of an Amazon Web Services account.

For a referenced security group in another VPC, the account ID of the referenced security group is returned in the response. If the referenced security group is deleted, this value is not returned.

EC2-Classic

Required when adding or removing rules that reference a security group in another Amazon Web Services account.

userIdGroupPair_peeringStatus :: Lens' UserIdGroupPair (Maybe Text)

The status of a VPC peering connection, if applicable.

userIdGroupPair_description :: Lens' UserIdGroupPair (Maybe Text)

A description for the security group rule that references this user ID group pair.

Constraints: Up to 255 characters in length. Allowed characters are a-z, A-Z, 0-9, spaces, and ._-:/()#,@[]+=;{}!$*

userIdGroupPair_vpcId :: Lens' UserIdGroupPair (Maybe Text)

The ID of the VPC for the referenced security group, if applicable.