Network.AWS.EC2.Types.VolumeModification

Description

Documentation

data VolumeModification

Describes the modification status of an EBS volume.

If the volume has never been modified, some element values will be null.

See: newVolumeModification smart constructor.

Constructors

VolumeModification'

Fields

statusMessage :: Maybe Text

A status message about the modification progress or failure.

originalVolumeType :: Maybe VolumeType

The original EBS volume type of the volume.

targetVolumeType :: Maybe VolumeType

The target EBS volume type of the volume.

startTime :: Maybe ISO8601

The modification start time.

volumeId :: Maybe Text

The ID of the volume.

endTime :: Maybe ISO8601

The modification completion or failure time.

targetIops :: Maybe Int

The target IOPS rate of the volume.

originalSize :: Maybe Int

The original size of the volume, in GiB.

targetSize :: Maybe Int

The target size of the volume, in GiB.

originalIops :: Maybe Int

The original IOPS rate of the volume.

modificationState :: Maybe VolumeModificationState

The current modification state. The modification state is null for unmodified volumes.

targetMultiAttachEnabled :: Maybe Bool

The target setting for Amazon EBS Multi-Attach.

originalMultiAttachEnabled :: Maybe Bool

The original setting for Amazon EBS Multi-Attach.

originalThroughput :: Maybe Int

The original throughput of the volume, in MiB/s.

targetThroughput :: Maybe Int

The target throughput of the volume, in MiB/s.

progress :: Maybe Integer

The modification progress, from 0 to 100 percent complete.

newVolumeModification :: VolumeModification

Create a value of VolumeModification with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:statusMessage:VolumeModification', volumeModification_statusMessage - A status message about the modification progress or failure.

$sel:originalVolumeType:VolumeModification', volumeModification_originalVolumeType - The original EBS volume type of the volume.

$sel:targetVolumeType:VolumeModification', volumeModification_targetVolumeType - The target EBS volume type of the volume.

$sel:startTime:VolumeModification', volumeModification_startTime - The modification start time.

$sel:volumeId:VolumeModification', volumeModification_volumeId - The ID of the volume.

$sel:endTime:VolumeModification', volumeModification_endTime - The modification completion or failure time.

$sel:targetIops:VolumeModification', volumeModification_targetIops - The target IOPS rate of the volume.

$sel:originalSize:VolumeModification', volumeModification_originalSize - The original size of the volume, in GiB.

$sel:targetSize:VolumeModification', volumeModification_targetSize - The target size of the volume, in GiB.

$sel:originalIops:VolumeModification', volumeModification_originalIops - The original IOPS rate of the volume.

$sel:modificationState:VolumeModification', volumeModification_modificationState - The current modification state. The modification state is null for unmodified volumes.

$sel:targetMultiAttachEnabled:VolumeModification', volumeModification_targetMultiAttachEnabled - The target setting for Amazon EBS Multi-Attach.

$sel:originalMultiAttachEnabled:VolumeModification', volumeModification_originalMultiAttachEnabled - The original setting for Amazon EBS Multi-Attach.

$sel:originalThroughput:VolumeModification', volumeModification_originalThroughput - The original throughput of the volume, in MiB/s.

$sel:targetThroughput:VolumeModification', volumeModification_targetThroughput - The target throughput of the volume, in MiB/s.

$sel:progress:VolumeModification', volumeModification_progress - The modification progress, from 0 to 100 percent complete.

volumeModification_statusMessage :: Lens' VolumeModification (Maybe Text)

A status message about the modification progress or failure.

volumeModification_originalVolumeType :: Lens' VolumeModification (Maybe VolumeType)

The original EBS volume type of the volume.

volumeModification_targetVolumeType :: Lens' VolumeModification (Maybe VolumeType)

The target EBS volume type of the volume.

volumeModification_startTime :: Lens' VolumeModification (Maybe UTCTime)

The modification start time.

volumeModification_volumeId :: Lens' VolumeModification (Maybe Text)

The ID of the volume.

volumeModification_endTime :: Lens' VolumeModification (Maybe UTCTime)

The modification completion or failure time.

volumeModification_targetIops :: Lens' VolumeModification (Maybe Int)

The target IOPS rate of the volume.

volumeModification_originalSize :: Lens' VolumeModification (Maybe Int)

The original size of the volume, in GiB.

volumeModification_targetSize :: Lens' VolumeModification (Maybe Int)

The target size of the volume, in GiB.

volumeModification_originalIops :: Lens' VolumeModification (Maybe Int)

The original IOPS rate of the volume.

volumeModification_modificationState :: Lens' VolumeModification (Maybe VolumeModificationState)

The current modification state. The modification state is null for unmodified volumes.

volumeModification_targetMultiAttachEnabled :: Lens' VolumeModification (Maybe Bool)

The target setting for Amazon EBS Multi-Attach.

volumeModification_originalMultiAttachEnabled :: Lens' VolumeModification (Maybe Bool)

The original setting for Amazon EBS Multi-Attach.

volumeModification_originalThroughput :: Lens' VolumeModification (Maybe Int)

The original throughput of the volume, in MiB/s.

volumeModification_targetThroughput :: Lens' VolumeModification (Maybe Int)

The target throughput of the volume, in MiB/s.

volumeModification_progress :: Lens' VolumeModification (Maybe Integer)

The modification progress, from 0 to 100 percent complete.