Network.AWS.EC2.Types.VpcIpv6CidrBlockAssociation

Description

Documentation

data VpcIpv6CidrBlockAssociation

Describes an IPv6 CIDR block associated with a VPC.

See: newVpcIpv6CidrBlockAssociation smart constructor.

Constructors

VpcIpv6CidrBlockAssociation'

Fields

ipv6CidrBlockState :: Maybe VpcCidrBlockState

Information about the state of the CIDR block.

ipv6Pool :: Maybe Text

The ID of the IPv6 address pool from which the IPv6 CIDR block is allocated.

ipv6CidrBlock :: Maybe Text

The IPv6 CIDR block.

associationId :: Maybe Text

The association ID for the IPv6 CIDR block.

networkBorderGroup :: Maybe Text

The name of the unique set of Availability Zones, Local Zones, or Wavelength Zones from which AWS advertises IP addresses, for example, us-east-1-wl1-bos-wlz-1.

newVpcIpv6CidrBlockAssociation :: VpcIpv6CidrBlockAssociation

Create a value of VpcIpv6CidrBlockAssociation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ipv6CidrBlockState:VpcIpv6CidrBlockAssociation', vpcIpv6CidrBlockAssociation_ipv6CidrBlockState - Information about the state of the CIDR block.

$sel:ipv6Pool:VpcIpv6CidrBlockAssociation', vpcIpv6CidrBlockAssociation_ipv6Pool - The ID of the IPv6 address pool from which the IPv6 CIDR block is allocated.

$sel:ipv6CidrBlock:VpcIpv6CidrBlockAssociation', vpcIpv6CidrBlockAssociation_ipv6CidrBlock - The IPv6 CIDR block.

$sel:associationId:VpcIpv6CidrBlockAssociation', vpcIpv6CidrBlockAssociation_associationId - The association ID for the IPv6 CIDR block.

$sel:networkBorderGroup:VpcIpv6CidrBlockAssociation', vpcIpv6CidrBlockAssociation_networkBorderGroup - The name of the unique set of Availability Zones, Local Zones, or Wavelength Zones from which AWS advertises IP addresses, for example, us-east-1-wl1-bos-wlz-1.

vpcIpv6CidrBlockAssociation_ipv6CidrBlockState :: Lens' VpcIpv6CidrBlockAssociation (Maybe VpcCidrBlockState)

Information about the state of the CIDR block.

vpcIpv6CidrBlockAssociation_ipv6Pool :: Lens' VpcIpv6CidrBlockAssociation (Maybe Text)

The ID of the IPv6 address pool from which the IPv6 CIDR block is allocated.

vpcIpv6CidrBlockAssociation_ipv6CidrBlock :: Lens' VpcIpv6CidrBlockAssociation (Maybe Text)

The IPv6 CIDR block.

vpcIpv6CidrBlockAssociation_associationId :: Lens' VpcIpv6CidrBlockAssociation (Maybe Text)

The association ID for the IPv6 CIDR block.

vpcIpv6CidrBlockAssociation_networkBorderGroup :: Lens' VpcIpv6CidrBlockAssociation (Maybe Text)

The name of the unique set of Availability Zones, Local Zones, or Wavelength Zones from which AWS advertises IP addresses, for example, us-east-1-wl1-bos-wlz-1.