Network.AWS.EC2.Types.VpcPeeringConnectionOptionsDescription

Description

Documentation

data VpcPeeringConnectionOptionsDescription

Describes the VPC peering connection options.

See: newVpcPeeringConnectionOptionsDescription smart constructor.

Constructors

VpcPeeringConnectionOptionsDescription'

Fields

allowEgressFromLocalVpcToRemoteClassicLink :: Maybe Bool

Indicates whether a local VPC can communicate with a ClassicLink connection in the peer VPC over the VPC peering connection.

allowEgressFromLocalClassicLinkToRemoteVpc :: Maybe Bool

Indicates whether a local ClassicLink connection can communicate with the peer VPC over the VPC peering connection.

allowDnsResolutionFromRemoteVpc :: Maybe Bool

Indicates whether a local VPC can resolve public DNS hostnames to private IP addresses when queried from instances in a peer VPC.

newVpcPeeringConnectionOptionsDescription :: VpcPeeringConnectionOptionsDescription

Create a value of VpcPeeringConnectionOptionsDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:allowEgressFromLocalVpcToRemoteClassicLink:VpcPeeringConnectionOptionsDescription', vpcPeeringConnectionOptionsDescription_allowEgressFromLocalVpcToRemoteClassicLink - Indicates whether a local VPC can communicate with a ClassicLink connection in the peer VPC over the VPC peering connection.

$sel:allowEgressFromLocalClassicLinkToRemoteVpc:VpcPeeringConnectionOptionsDescription', vpcPeeringConnectionOptionsDescription_allowEgressFromLocalClassicLinkToRemoteVpc - Indicates whether a local ClassicLink connection can communicate with the peer VPC over the VPC peering connection.

$sel:allowDnsResolutionFromRemoteVpc:VpcPeeringConnectionOptionsDescription', vpcPeeringConnectionOptionsDescription_allowDnsResolutionFromRemoteVpc - Indicates whether a local VPC can resolve public DNS hostnames to private IP addresses when queried from instances in a peer VPC.

vpcPeeringConnectionOptionsDescription_allowEgressFromLocalVpcToRemoteClassicLink :: Lens' VpcPeeringConnectionOptionsDescription (Maybe Bool)

Indicates whether a local VPC can communicate with a ClassicLink connection in the peer VPC over the VPC peering connection.

vpcPeeringConnectionOptionsDescription_allowEgressFromLocalClassicLinkToRemoteVpc :: Lens' VpcPeeringConnectionOptionsDescription (Maybe Bool)

Indicates whether a local ClassicLink connection can communicate with the peer VPC over the VPC peering connection.

vpcPeeringConnectionOptionsDescription_allowDnsResolutionFromRemoteVpc :: Lens' VpcPeeringConnectionOptionsDescription (Maybe Bool)

Indicates whether a local VPC can resolve public DNS hostnames to private IP addresses when queried from instances in a peer VPC.