Network.AWS.EC2.Types.VpcPeeringConnectionVpcInfo

Description

Documentation

data VpcPeeringConnectionVpcInfo

Describes a VPC in a VPC peering connection.

See: newVpcPeeringConnectionVpcInfo smart constructor.

Constructors

VpcPeeringConnectionVpcInfo'

Fields

cidrBlockSet :: Maybe [CidrBlock]

Information about the IPv4 CIDR blocks for the VPC.

vpcId :: Maybe Text

The ID of the VPC.

ownerId :: Maybe Text

The ID of the Amazon Web Services account that owns the VPC.

peeringOptions :: Maybe VpcPeeringConnectionOptionsDescription

Information about the VPC peering connection options for the accepter or requester VPC.

cidrBlock :: Maybe Text

The IPv4 CIDR block for the VPC.

region :: Maybe Text

The Region in which the VPC is located.

ipv6CidrBlockSet :: Maybe [Ipv6CidrBlock]

The IPv6 CIDR block for the VPC.

newVpcPeeringConnectionVpcInfo :: VpcPeeringConnectionVpcInfo

Create a value of VpcPeeringConnectionVpcInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cidrBlockSet:VpcPeeringConnectionVpcInfo', vpcPeeringConnectionVpcInfo_cidrBlockSet - Information about the IPv4 CIDR blocks for the VPC.

$sel:vpcId:VpcPeeringConnectionVpcInfo', vpcPeeringConnectionVpcInfo_vpcId - The ID of the VPC.

$sel:ownerId:VpcPeeringConnectionVpcInfo', vpcPeeringConnectionVpcInfo_ownerId - The ID of the Amazon Web Services account that owns the VPC.

$sel:peeringOptions:VpcPeeringConnectionVpcInfo', vpcPeeringConnectionVpcInfo_peeringOptions - Information about the VPC peering connection options for the accepter or requester VPC.

$sel:cidrBlock:VpcPeeringConnectionVpcInfo', vpcPeeringConnectionVpcInfo_cidrBlock - The IPv4 CIDR block for the VPC.

$sel:region:VpcPeeringConnectionVpcInfo', vpcPeeringConnectionVpcInfo_region - The Region in which the VPC is located.

$sel:ipv6CidrBlockSet:VpcPeeringConnectionVpcInfo', vpcPeeringConnectionVpcInfo_ipv6CidrBlockSet - The IPv6 CIDR block for the VPC.

vpcPeeringConnectionVpcInfo_cidrBlockSet :: Lens' VpcPeeringConnectionVpcInfo (Maybe [CidrBlock])

Information about the IPv4 CIDR blocks for the VPC.

vpcPeeringConnectionVpcInfo_vpcId :: Lens' VpcPeeringConnectionVpcInfo (Maybe Text)

The ID of the VPC.

vpcPeeringConnectionVpcInfo_ownerId :: Lens' VpcPeeringConnectionVpcInfo (Maybe Text)

The ID of the Amazon Web Services account that owns the VPC.

vpcPeeringConnectionVpcInfo_peeringOptions :: Lens' VpcPeeringConnectionVpcInfo (Maybe VpcPeeringConnectionOptionsDescription)

Information about the VPC peering connection options for the accepter or requester VPC.

vpcPeeringConnectionVpcInfo_cidrBlock :: Lens' VpcPeeringConnectionVpcInfo (Maybe Text)

The IPv4 CIDR block for the VPC.

vpcPeeringConnectionVpcInfo_region :: Lens' VpcPeeringConnectionVpcInfo (Maybe Text)

The Region in which the VPC is located.

vpcPeeringConnectionVpcInfo_ipv6CidrBlockSet :: Lens' VpcPeeringConnectionVpcInfo (Maybe [Ipv6CidrBlock])

The IPv6 CIDR block for the VPC.