Network.AWS.EC2.Types.VpnConnection

Description

Documentation

data VpnConnection

Describes a VPN connection.

See: newVpnConnection smart constructor.

Constructors

VpnConnection'

Fields

customerGatewayConfiguration :: Maybe Text

The configuration information for the VPN connection's customer gateway (in the native XML format). This element is always present in the CreateVpnConnection response; however, it's present in the DescribeVpnConnections response only if the VPN connection is in the pending or available state.

options :: Maybe VpnConnectionOptions

The VPN connection options.

category :: Maybe Text

The category of the VPN connection. A value of VPN indicates an Amazon Web Services VPN connection. A value of VPN-Classic indicates an Amazon Web Services Classic VPN connection.

routes :: Maybe [VpnStaticRoute]

The static routes associated with the VPN connection.

tags :: Maybe [Tag]

Any tags assigned to the VPN connection.

vgwTelemetry :: Maybe [VgwTelemetry]

Information about the VPN tunnel.

vpnGatewayId :: Maybe Text

The ID of the virtual private gateway at the Amazon Web Services side of the VPN connection.

transitGatewayId :: Maybe Text

The ID of the transit gateway associated with the VPN connection.

vpnConnectionId :: Text

The ID of the VPN connection.

customerGatewayId :: Text

The ID of the customer gateway at your end of the VPN connection.

state :: VpnState

The current state of the VPN connection.

type' :: GatewayType

The type of VPN connection.

newVpnConnection

Arguments

:: Text
$sel:vpnConnectionId:VpnConnection'

-> Text
$sel:customerGatewayId:VpnConnection'

-> VpnState
$sel:state:VpnConnection'

-> GatewayType
$sel:type':VpnConnection'

-> VpnConnection

Create a value of VpnConnection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:customerGatewayConfiguration:VpnConnection', vpnConnection_customerGatewayConfiguration - The configuration information for the VPN connection's customer gateway (in the native XML format). This element is always present in the CreateVpnConnection response; however, it's present in the DescribeVpnConnections response only if the VPN connection is in the pending or available state.

$sel:options:VpnConnection', vpnConnection_options - The VPN connection options.

$sel:category:VpnConnection', vpnConnection_category - The category of the VPN connection. A value of VPN indicates an Amazon Web Services VPN connection. A value of VPN-Classic indicates an Amazon Web Services Classic VPN connection.

$sel:routes:VpnConnection', vpnConnection_routes - The static routes associated with the VPN connection.

$sel:tags:VpnConnection', vpnConnection_tags - Any tags assigned to the VPN connection.

$sel:vgwTelemetry:VpnConnection', vpnConnection_vgwTelemetry - Information about the VPN tunnel.

$sel:vpnGatewayId:VpnConnection', vpnConnection_vpnGatewayId - The ID of the virtual private gateway at the Amazon Web Services side of the VPN connection.

$sel:transitGatewayId:VpnConnection', vpnConnection_transitGatewayId - The ID of the transit gateway associated with the VPN connection.

$sel:vpnConnectionId:VpnConnection', vpnConnection_vpnConnectionId - The ID of the VPN connection.

$sel:customerGatewayId:VpnConnection', vpnConnection_customerGatewayId - The ID of the customer gateway at your end of the VPN connection.

$sel:state:VpnConnection', vpnConnection_state - The current state of the VPN connection.

$sel:type':VpnConnection', vpnConnection_type - The type of VPN connection.

vpnConnection_customerGatewayConfiguration :: Lens' VpnConnection (Maybe Text)

The configuration information for the VPN connection's customer gateway (in the native XML format). This element is always present in the CreateVpnConnection response; however, it's present in the DescribeVpnConnections response only if the VPN connection is in the pending or available state.

vpnConnection_options :: Lens' VpnConnection (Maybe VpnConnectionOptions)

The VPN connection options.

vpnConnection_category :: Lens' VpnConnection (Maybe Text)

The category of the VPN connection. A value of VPN indicates an Amazon Web Services VPN connection. A value of VPN-Classic indicates an Amazon Web Services Classic VPN connection.

vpnConnection_routes :: Lens' VpnConnection (Maybe [VpnStaticRoute])

The static routes associated with the VPN connection.

vpnConnection_tags :: Lens' VpnConnection (Maybe [Tag])

Any tags assigned to the VPN connection.

vpnConnection_vgwTelemetry :: Lens' VpnConnection (Maybe [VgwTelemetry])

Information about the VPN tunnel.

vpnConnection_vpnGatewayId :: Lens' VpnConnection (Maybe Text)

The ID of the virtual private gateway at the Amazon Web Services side of the VPN connection.

vpnConnection_transitGatewayId :: Lens' VpnConnection (Maybe Text)

The ID of the transit gateway associated with the VPN connection.

vpnConnection_vpnConnectionId :: Lens' VpnConnection Text

The ID of the VPN connection.

vpnConnection_customerGatewayId :: Lens' VpnConnection Text

The ID of the customer gateway at your end of the VPN connection.

vpnConnection_state :: Lens' VpnConnection VpnState

The current state of the VPN connection.

vpnConnection_type :: Lens' VpnConnection GatewayType

The type of VPN connection.