Network.AWS.EC2.Types.VpnConnectionDeviceType

Description

Documentation

data VpnConnectionDeviceType

List of customer gateway devices that have a sample configuration file available for use. You can also see the list of device types with sample configuration files available under Your customer gateway device in the Amazon Web Services Site-to-Site VPN User Guide.

See: newVpnConnectionDeviceType smart constructor.

Constructors

VpnConnectionDeviceType'

Fields

platform :: Maybe Text

Customer gateway device platform.

vendor :: Maybe Text

Customer gateway device vendor.

software :: Maybe Text

Customer gateway device software version.

vpnConnectionDeviceTypeId :: Maybe Text

Customer gateway device identifier.

newVpnConnectionDeviceType :: VpnConnectionDeviceType

Create a value of VpnConnectionDeviceType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:platform:VpnConnectionDeviceType', vpnConnectionDeviceType_platform - Customer gateway device platform.

$sel:vendor:VpnConnectionDeviceType', vpnConnectionDeviceType_vendor - Customer gateway device vendor.

$sel:software:VpnConnectionDeviceType', vpnConnectionDeviceType_software - Customer gateway device software version.

$sel:vpnConnectionDeviceTypeId:VpnConnectionDeviceType', vpnConnectionDeviceType_vpnConnectionDeviceTypeId - Customer gateway device identifier.

vpnConnectionDeviceType_platform :: Lens' VpnConnectionDeviceType (Maybe Text)

Customer gateway device platform.

vpnConnectionDeviceType_vendor :: Lens' VpnConnectionDeviceType (Maybe Text)

Customer gateway device vendor.

vpnConnectionDeviceType_software :: Lens' VpnConnectionDeviceType (Maybe Text)

Customer gateway device software version.

vpnConnectionDeviceType_vpnConnectionDeviceTypeId :: Lens' VpnConnectionDeviceType (Maybe Text)

Customer gateway device identifier.