Network.AWS.EC2.Types.VpnConnectionOptions

Description

Documentation

data VpnConnectionOptions

Describes VPN connection options.

See: newVpnConnectionOptions smart constructor.

Constructors

VpnConnectionOptions'

Fields

remoteIpv6NetworkCidr :: Maybe Text

The IPv6 CIDR on the Amazon Web Services side of the VPN connection.

staticRoutesOnly :: Maybe Bool

Indicates whether the VPN connection uses static routes only. Static routes must be used for devices that don't support BGP.

localIpv6NetworkCidr :: Maybe Text

The IPv6 CIDR on the customer gateway (on-premises) side of the VPN connection.

enableAcceleration :: Maybe Bool

Indicates whether acceleration is enabled for the VPN connection.

tunnelOptions :: Maybe [TunnelOption]

Indicates the VPN tunnel options.

remoteIpv4NetworkCidr :: Maybe Text

The IPv4 CIDR on the Amazon Web Services side of the VPN connection.

tunnelInsideIpVersion :: Maybe TunnelInsideIpVersion

Indicates whether the VPN tunnels process IPv4 or IPv6 traffic.

localIpv4NetworkCidr :: Maybe Text

The IPv4 CIDR on the customer gateway (on-premises) side of the VPN connection.

newVpnConnectionOptions :: VpnConnectionOptions

Create a value of VpnConnectionOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:remoteIpv6NetworkCidr:VpnConnectionOptions', vpnConnectionOptions_remoteIpv6NetworkCidr - The IPv6 CIDR on the Amazon Web Services side of the VPN connection.

$sel:staticRoutesOnly:VpnConnectionOptions', vpnConnectionOptions_staticRoutesOnly - Indicates whether the VPN connection uses static routes only. Static routes must be used for devices that don't support BGP.

$sel:localIpv6NetworkCidr:VpnConnectionOptions', vpnConnectionOptions_localIpv6NetworkCidr - The IPv6 CIDR on the customer gateway (on-premises) side of the VPN connection.

$sel:enableAcceleration:VpnConnectionOptions', vpnConnectionOptions_enableAcceleration - Indicates whether acceleration is enabled for the VPN connection.

$sel:tunnelOptions:VpnConnectionOptions', vpnConnectionOptions_tunnelOptions - Indicates the VPN tunnel options.

$sel:remoteIpv4NetworkCidr:VpnConnectionOptions', vpnConnectionOptions_remoteIpv4NetworkCidr - The IPv4 CIDR on the Amazon Web Services side of the VPN connection.

$sel:tunnelInsideIpVersion:VpnConnectionOptions', vpnConnectionOptions_tunnelInsideIpVersion - Indicates whether the VPN tunnels process IPv4 or IPv6 traffic.

$sel:localIpv4NetworkCidr:VpnConnectionOptions', vpnConnectionOptions_localIpv4NetworkCidr - The IPv4 CIDR on the customer gateway (on-premises) side of the VPN connection.

vpnConnectionOptions_remoteIpv6NetworkCidr :: Lens' VpnConnectionOptions (Maybe Text)

The IPv6 CIDR on the Amazon Web Services side of the VPN connection.

vpnConnectionOptions_staticRoutesOnly :: Lens' VpnConnectionOptions (Maybe Bool)

Indicates whether the VPN connection uses static routes only. Static routes must be used for devices that don't support BGP.

vpnConnectionOptions_localIpv6NetworkCidr :: Lens' VpnConnectionOptions (Maybe Text)

The IPv6 CIDR on the customer gateway (on-premises) side of the VPN connection.

vpnConnectionOptions_enableAcceleration :: Lens' VpnConnectionOptions (Maybe Bool)

Indicates whether acceleration is enabled for the VPN connection.

vpnConnectionOptions_tunnelOptions :: Lens' VpnConnectionOptions (Maybe [TunnelOption])

Indicates the VPN tunnel options.

vpnConnectionOptions_remoteIpv4NetworkCidr :: Lens' VpnConnectionOptions (Maybe Text)

The IPv4 CIDR on the Amazon Web Services side of the VPN connection.

vpnConnectionOptions_tunnelInsideIpVersion :: Lens' VpnConnectionOptions (Maybe TunnelInsideIpVersion)

Indicates whether the VPN tunnels process IPv4 or IPv6 traffic.

vpnConnectionOptions_localIpv4NetworkCidr :: Lens' VpnConnectionOptions (Maybe Text)

The IPv4 CIDR on the customer gateway (on-premises) side of the VPN connection.