Network.AWS.EC2.Types.VpnConnectionOptionsSpecification

Description

Documentation

data VpnConnectionOptionsSpecification

Describes VPN connection options.

See: newVpnConnectionOptionsSpecification smart constructor.

Constructors

VpnConnectionOptionsSpecification'

Fields

remoteIpv6NetworkCidr :: Maybe Text

The IPv6 CIDR on the Amazon Web Services side of the VPN connection.

Default: ::/0

staticRoutesOnly :: Maybe Bool

Indicate whether the VPN connection uses static routes only. If you are creating a VPN connection for a device that does not support BGP, you must specify true. Use CreateVpnConnectionRoute to create a static route.

Default: false

localIpv6NetworkCidr :: Maybe Text

The IPv6 CIDR on the customer gateway (on-premises) side of the VPN connection.

Default: ::/0

enableAcceleration :: Maybe Bool

Indicate whether to enable acceleration for the VPN connection.

Default: false

tunnelOptions :: Maybe [VpnTunnelOptionsSpecification]

The tunnel options for the VPN connection.

remoteIpv4NetworkCidr :: Maybe Text

The IPv4 CIDR on the Amazon Web Services side of the VPN connection.

Default: 0.0.0.0/0

tunnelInsideIpVersion :: Maybe TunnelInsideIpVersion

Indicate whether the VPN tunnels process IPv4 or IPv6 traffic.

Default: ipv4

localIpv4NetworkCidr :: Maybe Text

The IPv4 CIDR on the customer gateway (on-premises) side of the VPN connection.

Default: 0.0.0.0/0

newVpnConnectionOptionsSpecification :: VpnConnectionOptionsSpecification

Create a value of VpnConnectionOptionsSpecification with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:remoteIpv6NetworkCidr:VpnConnectionOptionsSpecification', vpnConnectionOptionsSpecification_remoteIpv6NetworkCidr - The IPv6 CIDR on the Amazon Web Services side of the VPN connection.

Default: ::/0

$sel:staticRoutesOnly:VpnConnectionOptionsSpecification', vpnConnectionOptionsSpecification_staticRoutesOnly - Indicate whether the VPN connection uses static routes only. If you are creating a VPN connection for a device that does not support BGP, you must specify true. Use CreateVpnConnectionRoute to create a static route.

Default: false

$sel:localIpv6NetworkCidr:VpnConnectionOptionsSpecification', vpnConnectionOptionsSpecification_localIpv6NetworkCidr - The IPv6 CIDR on the customer gateway (on-premises) side of the VPN connection.

Default: ::/0

$sel:enableAcceleration:VpnConnectionOptionsSpecification', vpnConnectionOptionsSpecification_enableAcceleration - Indicate whether to enable acceleration for the VPN connection.

Default: false

$sel:tunnelOptions:VpnConnectionOptionsSpecification', vpnConnectionOptionsSpecification_tunnelOptions - The tunnel options for the VPN connection.

$sel:remoteIpv4NetworkCidr:VpnConnectionOptionsSpecification', vpnConnectionOptionsSpecification_remoteIpv4NetworkCidr - The IPv4 CIDR on the Amazon Web Services side of the VPN connection.

Default: 0.0.0.0/0

$sel:tunnelInsideIpVersion:VpnConnectionOptionsSpecification', vpnConnectionOptionsSpecification_tunnelInsideIpVersion - Indicate whether the VPN tunnels process IPv4 or IPv6 traffic.

Default: ipv4

$sel:localIpv4NetworkCidr:VpnConnectionOptionsSpecification', vpnConnectionOptionsSpecification_localIpv4NetworkCidr - The IPv4 CIDR on the customer gateway (on-premises) side of the VPN connection.

Default: 0.0.0.0/0

vpnConnectionOptionsSpecification_remoteIpv6NetworkCidr :: Lens' VpnConnectionOptionsSpecification (Maybe Text)

The IPv6 CIDR on the Amazon Web Services side of the VPN connection.

Default: ::/0

vpnConnectionOptionsSpecification_staticRoutesOnly :: Lens' VpnConnectionOptionsSpecification (Maybe Bool)

Indicate whether the VPN connection uses static routes only. If you are creating a VPN connection for a device that does not support BGP, you must specify true. Use CreateVpnConnectionRoute to create a static route.

Default: false

vpnConnectionOptionsSpecification_localIpv6NetworkCidr :: Lens' VpnConnectionOptionsSpecification (Maybe Text)

The IPv6 CIDR on the customer gateway (on-premises) side of the VPN connection.

Default: ::/0

vpnConnectionOptionsSpecification_enableAcceleration :: Lens' VpnConnectionOptionsSpecification (Maybe Bool)

Indicate whether to enable acceleration for the VPN connection.

Default: false

vpnConnectionOptionsSpecification_tunnelOptions :: Lens' VpnConnectionOptionsSpecification (Maybe [VpnTunnelOptionsSpecification])

The tunnel options for the VPN connection.

vpnConnectionOptionsSpecification_remoteIpv4NetworkCidr :: Lens' VpnConnectionOptionsSpecification (Maybe Text)

The IPv4 CIDR on the Amazon Web Services side of the VPN connection.

Default: 0.0.0.0/0

vpnConnectionOptionsSpecification_tunnelInsideIpVersion :: Lens' VpnConnectionOptionsSpecification (Maybe TunnelInsideIpVersion)

Indicate whether the VPN tunnels process IPv4 or IPv6 traffic.

Default: ipv4

vpnConnectionOptionsSpecification_localIpv4NetworkCidr :: Lens' VpnConnectionOptionsSpecification (Maybe Text)

The IPv4 CIDR on the customer gateway (on-premises) side of the VPN connection.

Default: 0.0.0.0/0