Network.AWS.EC2.Types.VpnGateway

Description

Documentation

data VpnGateway

Describes a virtual private gateway.

See: newVpnGateway smart constructor.

Constructors

VpnGateway'

Fields

vpcAttachments :: Maybe [VpcAttachment]

Any VPCs attached to the virtual private gateway.

availabilityZone :: Maybe Text

The Availability Zone where the virtual private gateway was created, if applicable. This field may be empty or not returned.

state :: Maybe VpnState

The current state of the virtual private gateway.

tags :: Maybe [Tag]

Any tags assigned to the virtual private gateway.

amazonSideAsn :: Maybe Integer

The private Autonomous System Number (ASN) for the Amazon side of a BGP session.

type' :: Maybe GatewayType

The type of VPN connection the virtual private gateway supports.

vpnGatewayId :: Maybe Text

The ID of the virtual private gateway.

newVpnGateway :: VpnGateway

Create a value of VpnGateway with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpcAttachments:VpnGateway', vpnGateway_vpcAttachments - Any VPCs attached to the virtual private gateway.

$sel:availabilityZone:VpnGateway', vpnGateway_availabilityZone - The Availability Zone where the virtual private gateway was created, if applicable. This field may be empty or not returned.

$sel:state:VpnGateway', vpnGateway_state - The current state of the virtual private gateway.

$sel:tags:VpnGateway', vpnGateway_tags - Any tags assigned to the virtual private gateway.

$sel:amazonSideAsn:VpnGateway', vpnGateway_amazonSideAsn - The private Autonomous System Number (ASN) for the Amazon side of a BGP session.

$sel:type':VpnGateway', vpnGateway_type - The type of VPN connection the virtual private gateway supports.

$sel:vpnGatewayId:VpnGateway', vpnGateway_vpnGatewayId - The ID of the virtual private gateway.

vpnGateway_vpcAttachments :: Lens' VpnGateway (Maybe [VpcAttachment])

Any VPCs attached to the virtual private gateway.

vpnGateway_availabilityZone :: Lens' VpnGateway (Maybe Text)

The Availability Zone where the virtual private gateway was created, if applicable. This field may be empty or not returned.

vpnGateway_state :: Lens' VpnGateway (Maybe VpnState)

The current state of the virtual private gateway.

vpnGateway_tags :: Lens' VpnGateway (Maybe [Tag])

Any tags assigned to the virtual private gateway.

vpnGateway_amazonSideAsn :: Lens' VpnGateway (Maybe Integer)

The private Autonomous System Number (ASN) for the Amazon side of a BGP session.

vpnGateway_type :: Lens' VpnGateway (Maybe GatewayType)

The type of VPN connection the virtual private gateway supports.

vpnGateway_vpnGatewayId :: Lens' VpnGateway (Maybe Text)

The ID of the virtual private gateway.