Network.AWS.EC2.Types.VpnTunnelOptionsSpecification

Description

Documentation

data VpnTunnelOptionsSpecification

The tunnel options for a single VPN tunnel.

See: newVpnTunnelOptionsSpecification smart constructor.

Constructors

VpnTunnelOptionsSpecification'

Fields

phase1LifetimeSeconds :: Maybe Int

The lifetime for phase 1 of the IKE negotiation, in seconds.

Constraints: A value between 900 and 28,800.

Default: 28800

iKEVersions :: Maybe [IKEVersionsRequestListValue]

The IKE versions that are permitted for the VPN tunnel.

Valid values: ikev1 | ikev2

phase2IntegrityAlgorithms :: Maybe [Phase2IntegrityAlgorithmsRequestListValue]

One or more integrity algorithms that are permitted for the VPN tunnel for phase 2 IKE negotiations.

Valid values: SHA1 | SHA2-256 | SHA2-384 | SHA2-512

phase2EncryptionAlgorithms :: Maybe [Phase2EncryptionAlgorithmsRequestListValue]

One or more encryption algorithms that are permitted for the VPN tunnel for phase 2 IKE negotiations.

Valid values: AES128 | AES256 | AES128-GCM-16 | AES256-GCM-16

phase2DHGroupNumbers :: Maybe [Phase2DHGroupNumbersRequestListValue]

One or more Diffie-Hellman group numbers that are permitted for the VPN tunnel for phase 2 IKE negotiations.

Valid values: 2 | 5 | 14 | 15 | 16 | 17 | 18 | 19 | 20 | 21 | 22 | 23 | 24

startupAction :: Maybe Text

The action to take when the establishing the tunnel for the VPN connection. By default, your customer gateway device must initiate the IKE negotiation and bring up the tunnel. Specify start for Amazon Web Services to initiate the IKE negotiation.

Valid Values: add | start

Default: add

dPDTimeoutSeconds :: Maybe Int

The number of seconds after which a DPD timeout occurs.

Constraints: A value between 0 and 30.

Default: 30

phase1DHGroupNumbers :: Maybe [Phase1DHGroupNumbersRequestListValue]

One or more Diffie-Hellman group numbers that are permitted for the VPN tunnel for phase 1 IKE negotiations.

Valid values: 2 | 14 | 15 | 16 | 17 | 18 | 19 | 20 | 21 | 22 | 23 | 24

phase1EncryptionAlgorithms :: Maybe [Phase1EncryptionAlgorithmsRequestListValue]

One or more encryption algorithms that are permitted for the VPN tunnel for phase 1 IKE negotiations.

Valid values: AES128 | AES256 | AES128-GCM-16 | AES256-GCM-16

replayWindowSize :: Maybe Int

The number of packets in an IKE replay window.

Constraints: A value between 64 and 2048.

Default: 1024

preSharedKey :: Maybe Text

The pre-shared key (PSK) to establish initial authentication between the virtual private gateway and customer gateway.

Constraints: Allowed characters are alphanumeric characters, periods (.), and underscores (_). Must be between 8 and 64 characters in length and cannot start with zero (0).

tunnelInsideIpv6Cidr :: Maybe Text

The range of inside IPv6 addresses for the tunnel. Any specified CIDR blocks must be unique across all VPN connections that use the same transit gateway.

Constraints: A size /126 CIDR block from the local fd00::/8 range.

rekeyFuzzPercentage :: Maybe Int

The percentage of the rekey window (determined by RekeyMarginTimeSeconds) during which the rekey time is randomly selected.

Constraints: A value between 0 and 100.

Default: 100

rekeyMarginTimeSeconds :: Maybe Int

The margin time, in seconds, before the phase 2 lifetime expires, during which the Amazon Web Services side of the VPN connection performs an IKE rekey. The exact time of the rekey is randomly selected based on the value for RekeyFuzzPercentage.

Constraints: A value between 60 and half of Phase2LifetimeSeconds.

Default: 540

tunnelInsideCidr :: Maybe Text

The range of inside IPv4 addresses for the tunnel. Any specified CIDR blocks must be unique across all VPN connections that use the same virtual private gateway.

Constraints: A size /30 CIDR block from the 169.254.0.0/16 range. The following CIDR blocks are reserved and cannot be used:

169.254.0.0/30

169.254.1.0/30

169.254.2.0/30

169.254.3.0/30

169.254.4.0/30

169.254.5.0/30

169.254.169.252/30

phase1IntegrityAlgorithms :: Maybe [Phase1IntegrityAlgorithmsRequestListValue]

One or more integrity algorithms that are permitted for the VPN tunnel for phase 1 IKE negotiations.

Valid values: SHA1 | SHA2-256 | SHA2-384 | SHA2-512

dPDTimeoutAction :: Maybe Text

The action to take after DPD timeout occurs. Specify restart to restart the IKE initiation. Specify clear to end the IKE session.

Valid Values: clear | none | restart

Default: clear

phase2LifetimeSeconds :: Maybe Int

The lifetime for phase 2 of the IKE negotiation, in seconds.

Constraints: A value between 900 and 3,600. The value must be less than the value for Phase1LifetimeSeconds.

Default: 3600

newVpnTunnelOptionsSpecification :: VpnTunnelOptionsSpecification

Create a value of VpnTunnelOptionsSpecification with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:phase1LifetimeSeconds:VpnTunnelOptionsSpecification', vpnTunnelOptionsSpecification_phase1LifetimeSeconds - The lifetime for phase 1 of the IKE negotiation, in seconds.

Constraints: A value between 900 and 28,800.

Default: 28800

$sel:iKEVersions:VpnTunnelOptionsSpecification', vpnTunnelOptionsSpecification_iKEVersions - The IKE versions that are permitted for the VPN tunnel.

Valid values: ikev1 | ikev2

$sel:phase2IntegrityAlgorithms:VpnTunnelOptionsSpecification', vpnTunnelOptionsSpecification_phase2IntegrityAlgorithms - One or more integrity algorithms that are permitted for the VPN tunnel for phase 2 IKE negotiations.

Valid values: SHA1 | SHA2-256 | SHA2-384 | SHA2-512

$sel:phase2EncryptionAlgorithms:VpnTunnelOptionsSpecification', vpnTunnelOptionsSpecification_phase2EncryptionAlgorithms - One or more encryption algorithms that are permitted for the VPN tunnel for phase 2 IKE negotiations.

Valid values: AES128 | AES256 | AES128-GCM-16 | AES256-GCM-16

$sel:phase2DHGroupNumbers:VpnTunnelOptionsSpecification', vpnTunnelOptionsSpecification_phase2DHGroupNumbers - One or more Diffie-Hellman group numbers that are permitted for the VPN tunnel for phase 2 IKE negotiations.

Valid values: 2 | 5 | 14 | 15 | 16 | 17 | 18 | 19 | 20 | 21 | 22 | 23 | 24

$sel:startupAction:VpnTunnelOptionsSpecification', vpnTunnelOptionsSpecification_startupAction - The action to take when the establishing the tunnel for the VPN connection. By default, your customer gateway device must initiate the IKE negotiation and bring up the tunnel. Specify start for Amazon Web Services to initiate the IKE negotiation.

Valid Values: add | start

Default: add

$sel:dPDTimeoutSeconds:VpnTunnelOptionsSpecification', vpnTunnelOptionsSpecification_dPDTimeoutSeconds - The number of seconds after which a DPD timeout occurs.

Constraints: A value between 0 and 30.

Default: 30

$sel:phase1DHGroupNumbers:VpnTunnelOptionsSpecification', vpnTunnelOptionsSpecification_phase1DHGroupNumbers - One or more Diffie-Hellman group numbers that are permitted for the VPN tunnel for phase 1 IKE negotiations.

Valid values: 2 | 14 | 15 | 16 | 17 | 18 | 19 | 20 | 21 | 22 | 23 | 24

$sel:phase1EncryptionAlgorithms:VpnTunnelOptionsSpecification', vpnTunnelOptionsSpecification_phase1EncryptionAlgorithms - One or more encryption algorithms that are permitted for the VPN tunnel for phase 1 IKE negotiations.

Valid values: AES128 | AES256 | AES128-GCM-16 | AES256-GCM-16

$sel:replayWindowSize:VpnTunnelOptionsSpecification', vpnTunnelOptionsSpecification_replayWindowSize - The number of packets in an IKE replay window.

Constraints: A value between 64 and 2048.

Default: 1024

$sel:preSharedKey:VpnTunnelOptionsSpecification', vpnTunnelOptionsSpecification_preSharedKey - The pre-shared key (PSK) to establish initial authentication between the virtual private gateway and customer gateway.

Constraints: Allowed characters are alphanumeric characters, periods (.), and underscores (_). Must be between 8 and 64 characters in length and cannot start with zero (0).

$sel:tunnelInsideIpv6Cidr:VpnTunnelOptionsSpecification', vpnTunnelOptionsSpecification_tunnelInsideIpv6Cidr - The range of inside IPv6 addresses for the tunnel. Any specified CIDR blocks must be unique across all VPN connections that use the same transit gateway.

Constraints: A size /126 CIDR block from the local fd00::/8 range.

$sel:rekeyFuzzPercentage:VpnTunnelOptionsSpecification', vpnTunnelOptionsSpecification_rekeyFuzzPercentage - The percentage of the rekey window (determined by RekeyMarginTimeSeconds) during which the rekey time is randomly selected.

Constraints: A value between 0 and 100.

Default: 100

$sel:rekeyMarginTimeSeconds:VpnTunnelOptionsSpecification', vpnTunnelOptionsSpecification_rekeyMarginTimeSeconds - The margin time, in seconds, before the phase 2 lifetime expires, during which the Amazon Web Services side of the VPN connection performs an IKE rekey. The exact time of the rekey is randomly selected based on the value for RekeyFuzzPercentage.

Constraints: A value between 60 and half of Phase2LifetimeSeconds.

Default: 540

$sel:tunnelInsideCidr:VpnTunnelOptionsSpecification', vpnTunnelOptionsSpecification_tunnelInsideCidr - The range of inside IPv4 addresses for the tunnel. Any specified CIDR blocks must be unique across all VPN connections that use the same virtual private gateway.

Constraints: A size /30 CIDR block from the 169.254.0.0/16 range. The following CIDR blocks are reserved and cannot be used:

169.254.0.0/30

169.254.1.0/30

169.254.2.0/30

169.254.3.0/30

169.254.4.0/30

169.254.5.0/30

169.254.169.252/30

$sel:phase1IntegrityAlgorithms:VpnTunnelOptionsSpecification', vpnTunnelOptionsSpecification_phase1IntegrityAlgorithms - One or more integrity algorithms that are permitted for the VPN tunnel for phase 1 IKE negotiations.

Valid values: SHA1 | SHA2-256 | SHA2-384 | SHA2-512

$sel:dPDTimeoutAction:VpnTunnelOptionsSpecification', vpnTunnelOptionsSpecification_dPDTimeoutAction - The action to take after DPD timeout occurs. Specify restart to restart the IKE initiation. Specify clear to end the IKE session.

Valid Values: clear | none | restart

Default: clear

$sel:phase2LifetimeSeconds:VpnTunnelOptionsSpecification', vpnTunnelOptionsSpecification_phase2LifetimeSeconds - The lifetime for phase 2 of the IKE negotiation, in seconds.

Constraints: A value between 900 and 3,600. The value must be less than the value for Phase1LifetimeSeconds.

Default: 3600

vpnTunnelOptionsSpecification_phase1LifetimeSeconds :: Lens' VpnTunnelOptionsSpecification (Maybe Int)

The lifetime for phase 1 of the IKE negotiation, in seconds.

Constraints: A value between 900 and 28,800.

Default: 28800

vpnTunnelOptionsSpecification_iKEVersions :: Lens' VpnTunnelOptionsSpecification (Maybe [IKEVersionsRequestListValue])

The IKE versions that are permitted for the VPN tunnel.

Valid values: ikev1 | ikev2

vpnTunnelOptionsSpecification_phase2IntegrityAlgorithms :: Lens' VpnTunnelOptionsSpecification (Maybe [Phase2IntegrityAlgorithmsRequestListValue])

One or more integrity algorithms that are permitted for the VPN tunnel for phase 2 IKE negotiations.

Valid values: SHA1 | SHA2-256 | SHA2-384 | SHA2-512

vpnTunnelOptionsSpecification_phase2EncryptionAlgorithms :: Lens' VpnTunnelOptionsSpecification (Maybe [Phase2EncryptionAlgorithmsRequestListValue])

One or more encryption algorithms that are permitted for the VPN tunnel for phase 2 IKE negotiations.

Valid values: AES128 | AES256 | AES128-GCM-16 | AES256-GCM-16

vpnTunnelOptionsSpecification_phase2DHGroupNumbers :: Lens' VpnTunnelOptionsSpecification (Maybe [Phase2DHGroupNumbersRequestListValue])

One or more Diffie-Hellman group numbers that are permitted for the VPN tunnel for phase 2 IKE negotiations.

Valid values: 2 | 5 | 14 | 15 | 16 | 17 | 18 | 19 | 20 | 21 | 22 | 23 | 24

vpnTunnelOptionsSpecification_startupAction :: Lens' VpnTunnelOptionsSpecification (Maybe Text)

The action to take when the establishing the tunnel for the VPN connection. By default, your customer gateway device must initiate the IKE negotiation and bring up the tunnel. Specify start for Amazon Web Services to initiate the IKE negotiation.

Valid Values: add | start

Default: add

vpnTunnelOptionsSpecification_dPDTimeoutSeconds :: Lens' VpnTunnelOptionsSpecification (Maybe Int)

The number of seconds after which a DPD timeout occurs.

Constraints: A value between 0 and 30.

Default: 30

vpnTunnelOptionsSpecification_phase1DHGroupNumbers :: Lens' VpnTunnelOptionsSpecification (Maybe [Phase1DHGroupNumbersRequestListValue])

One or more Diffie-Hellman group numbers that are permitted for the VPN tunnel for phase 1 IKE negotiations.

Valid values: 2 | 14 | 15 | 16 | 17 | 18 | 19 | 20 | 21 | 22 | 23 | 24

vpnTunnelOptionsSpecification_phase1EncryptionAlgorithms :: Lens' VpnTunnelOptionsSpecification (Maybe [Phase1EncryptionAlgorithmsRequestListValue])

One or more encryption algorithms that are permitted for the VPN tunnel for phase 1 IKE negotiations.

Valid values: AES128 | AES256 | AES128-GCM-16 | AES256-GCM-16

vpnTunnelOptionsSpecification_replayWindowSize :: Lens' VpnTunnelOptionsSpecification (Maybe Int)

The number of packets in an IKE replay window.

Constraints: A value between 64 and 2048.

Default: 1024

vpnTunnelOptionsSpecification_preSharedKey :: Lens' VpnTunnelOptionsSpecification (Maybe Text)

The pre-shared key (PSK) to establish initial authentication between the virtual private gateway and customer gateway.

Constraints: Allowed characters are alphanumeric characters, periods (.), and underscores (_). Must be between 8 and 64 characters in length and cannot start with zero (0).

vpnTunnelOptionsSpecification_tunnelInsideIpv6Cidr :: Lens' VpnTunnelOptionsSpecification (Maybe Text)

The range of inside IPv6 addresses for the tunnel. Any specified CIDR blocks must be unique across all VPN connections that use the same transit gateway.

Constraints: A size /126 CIDR block from the local fd00::/8 range.

vpnTunnelOptionsSpecification_rekeyFuzzPercentage :: Lens' VpnTunnelOptionsSpecification (Maybe Int)

The percentage of the rekey window (determined by RekeyMarginTimeSeconds) during which the rekey time is randomly selected.

Constraints: A value between 0 and 100.

Default: 100

vpnTunnelOptionsSpecification_rekeyMarginTimeSeconds :: Lens' VpnTunnelOptionsSpecification (Maybe Int)

The margin time, in seconds, before the phase 2 lifetime expires, during which the Amazon Web Services side of the VPN connection performs an IKE rekey. The exact time of the rekey is randomly selected based on the value for RekeyFuzzPercentage.

Constraints: A value between 60 and half of Phase2LifetimeSeconds.

Default: 540

vpnTunnelOptionsSpecification_tunnelInsideCidr :: Lens' VpnTunnelOptionsSpecification (Maybe Text)

The range of inside IPv4 addresses for the tunnel. Any specified CIDR blocks must be unique across all VPN connections that use the same virtual private gateway.

Constraints: A size /30 CIDR block from the 169.254.0.0/16 range. The following CIDR blocks are reserved and cannot be used:

169.254.0.0/30

169.254.1.0/30

169.254.2.0/30

169.254.3.0/30

169.254.4.0/30

169.254.5.0/30

169.254.169.252/30

vpnTunnelOptionsSpecification_phase1IntegrityAlgorithms :: Lens' VpnTunnelOptionsSpecification (Maybe [Phase1IntegrityAlgorithmsRequestListValue])

One or more integrity algorithms that are permitted for the VPN tunnel for phase 1 IKE negotiations.

Valid values: SHA1 | SHA2-256 | SHA2-384 | SHA2-512

vpnTunnelOptionsSpecification_dPDTimeoutAction :: Lens' VpnTunnelOptionsSpecification (Maybe Text)

The action to take after DPD timeout occurs. Specify restart to restart the IKE initiation. Specify clear to end the IKE session.

Valid Values: clear | none | restart

Default: clear

vpnTunnelOptionsSpecification_phase2LifetimeSeconds :: Lens' VpnTunnelOptionsSpecification (Maybe Int)

The lifetime for phase 2 of the IKE negotiation, in seconds.

Constraints: A value between 900 and 3,600. The value must be less than the value for Phase1LifetimeSeconds.

Default: 3600