Network.AWS.EC2.UpdateSecurityGroupRuleDescriptionsEgress

Description

VPC only

Updates the description of an egress (outbound) security group rule. You can replace an existing description, or add a description to a rule that did not have one previously. You can remove a description for a security group rule by omitting the description parameter in the request.

Creating a Request

data UpdateSecurityGroupRuleDescriptionsEgress

See: newUpdateSecurityGroupRuleDescriptionsEgress smart constructor.

Constructors

UpdateSecurityGroupRuleDescriptionsEgress'

Fields

securityGroupRuleDescriptions :: Maybe [SecurityGroupRuleDescription]

The description for the egress security group rules. You must specify either the description or the IP permissions.

ipPermissions :: Maybe [IpPermission]

The IP permissions for the security group rule. You must specify either the IP permissions or the description.

groupId :: Maybe Text

The ID of the security group. You must specify either the security group ID or the security group name in the request. For security groups in a nondefault VPC, you must specify the security group ID.

groupName :: Maybe Text

Default VPC

The name of the security group. You must specify either the security group ID or the security group name in the request.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

newUpdateSecurityGroupRuleDescriptionsEgress :: UpdateSecurityGroupRuleDescriptionsEgress

Create a value of UpdateSecurityGroupRuleDescriptionsEgress with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:securityGroupRuleDescriptions:UpdateSecurityGroupRuleDescriptionsEgress', updateSecurityGroupRuleDescriptionsEgress_securityGroupRuleDescriptions - The description for the egress security group rules. You must specify either the description or the IP permissions.

$sel:ipPermissions:UpdateSecurityGroupRuleDescriptionsEgress', updateSecurityGroupRuleDescriptionsEgress_ipPermissions - The IP permissions for the security group rule. You must specify either the IP permissions or the description.

$sel:groupId:UpdateSecurityGroupRuleDescriptionsEgress', updateSecurityGroupRuleDescriptionsEgress_groupId - The ID of the security group. You must specify either the security group ID or the security group name in the request. For security groups in a nondefault VPC, you must specify the security group ID.

$sel:groupName:UpdateSecurityGroupRuleDescriptionsEgress', updateSecurityGroupRuleDescriptionsEgress_groupName - [Default VPC] The name of the security group. You must specify either the security group ID or the security group name in the request.

$sel:dryRun:UpdateSecurityGroupRuleDescriptionsEgress', updateSecurityGroupRuleDescriptionsEgress_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

Request Lenses

updateSecurityGroupRuleDescriptionsEgress_securityGroupRuleDescriptions :: Lens' UpdateSecurityGroupRuleDescriptionsEgress (Maybe [SecurityGroupRuleDescription])

The description for the egress security group rules. You must specify either the description or the IP permissions.

updateSecurityGroupRuleDescriptionsEgress_ipPermissions :: Lens' UpdateSecurityGroupRuleDescriptionsEgress (Maybe [IpPermission])

The IP permissions for the security group rule. You must specify either the IP permissions or the description.

updateSecurityGroupRuleDescriptionsEgress_groupId :: Lens' UpdateSecurityGroupRuleDescriptionsEgress (Maybe Text)

The ID of the security group. You must specify either the security group ID or the security group name in the request. For security groups in a nondefault VPC, you must specify the security group ID.

updateSecurityGroupRuleDescriptionsEgress_groupName :: Lens' UpdateSecurityGroupRuleDescriptionsEgress (Maybe Text)

Default VPC

The name of the security group. You must specify either the security group ID or the security group name in the request.

updateSecurityGroupRuleDescriptionsEgress_dryRun :: Lens' UpdateSecurityGroupRuleDescriptionsEgress (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

Destructuring the Response

data UpdateSecurityGroupRuleDescriptionsEgressResponse

See: newUpdateSecurityGroupRuleDescriptionsEgressResponse smart constructor.

Constructors

UpdateSecurityGroupRuleDescriptionsEgressResponse'

Fields

return' :: Maybe Bool

Returns true if the request succeeds; otherwise, returns an error.

httpStatus :: Int

The response's http status code.

newUpdateSecurityGroupRuleDescriptionsEgressResponse

Arguments

:: Int
$sel:httpStatus:UpdateSecurityGroupRuleDescriptionsEgressResponse'

-> UpdateSecurityGroupRuleDescriptionsEgressResponse

Create a value of UpdateSecurityGroupRuleDescriptionsEgressResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:return':UpdateSecurityGroupRuleDescriptionsEgressResponse', updateSecurityGroupRuleDescriptionsEgressResponse_return - Returns true if the request succeeds; otherwise, returns an error.

$sel:httpStatus:UpdateSecurityGroupRuleDescriptionsEgressResponse', updateSecurityGroupRuleDescriptionsEgressResponse_httpStatus - The response's http status code.

Response Lenses

updateSecurityGroupRuleDescriptionsEgressResponse_return :: Lens' UpdateSecurityGroupRuleDescriptionsEgressResponse (Maybe Bool)

Returns true if the request succeeds; otherwise, returns an error.

updateSecurityGroupRuleDescriptionsEgressResponse_httpStatus :: Lens' UpdateSecurityGroupRuleDescriptionsEgressResponse Int

The response's http status code.