Network.AWS.EC2.V2014_06_15.CreateImage

Description

Creates an Amazon EBS-backed AMI from an Amazon EBS-backed instance that is either running or stopped. If you customized your instance with instance store volumes or EBS volumes in addition to the root device volume, the new AMI contains block device mapping information for those volumes. When you launch an instance from this new AMI, the instance automatically launches with those additional volumes. For more information, see Creating Amazon EBS-Backed Linux AMIs in the Amazon Elastic Compute Cloud User Guide. Example This example request creates an AMI from the specified instance. https:/ec2.amazonaws.com?Action=CreateImage &amp;Description=Standard+Web+Server+v1.0 &amp;InstanceId=i-10a64379 &amp;Name=standard-web-server-v1.0 &amp;AUTHPARAMS &lt;CreateImageResponse xmlns="http://ec2.amazonaws.com/doc/2014-02-01/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;imageId&gt;ami-4fa54026&lt;imageId&gt; &lt;CreateImageResponse&gt;.

Documentation

createImage

Arguments

:: Text
_citInstanceId

-> Text
_citName

-> CreateImage

Minimum specification for a CreateImage request.

data CreateImage

Constructors

CreateImage

Fields

_citInstanceId :: Text

The ID of the instance.

_citName :: Text

A name for the new image. Constraints: 3-128 alphanumeric characters, parenthesis (()), periods (.), slashes (/), dashes (-), or underscores(_).

_citBlockDeviceMappings :: [BlockDeviceMapping]

Information about one or more block device mappings.

_citNoReboot :: Maybe Bool

By default, this parameter is set to false, which means Amazon EC2 attempts to shut down the instance cleanly before image creation and then reboots the instance. When the parameter is set to true, Amazon EC2 doesn't shut down the instance before creating the image. When this option is used, file system integrity on the created image can't be guaranteed.

_citDryRun :: Maybe Bool

_citDescription :: Maybe Text

A description for the new image.

citNoReboot :: Lens' CreateImage (Maybe Bool)

citName :: Lens' CreateImage Text

citInstanceId :: Lens' CreateImage Text

citDryRun :: Lens' CreateImage (Maybe Bool)

citDescription :: Lens' CreateImage (Maybe Text)

citBlockDeviceMappings :: Lens' CreateImage [BlockDeviceMapping]

data CreateImageResponse

Constructors

CreateImageResponse

Fields

_ciuImageId :: Maybe Text

The ID of the new AMI.

ciuImageId :: Iso' CreateImageResponse (Maybe Text)