Network.AWS.EC2.V2014_06_15.CreateInternetGateway

Description

Creates an Internet gateway for use with a VPC. After creating the Internet gateway, you attach it to a VPC using AttachInternetGateway. For more information about your VPC and Internet gateway, see the Amazon Virtual Private Cloud User Guide. Example This example creates an Internet gateway. https:/ec2.amazonaws.com?Action=CreateInternetGateway &amp;AUTHPARAMS 59dbff89-35bd-4eac-99ed-be587EXAMPLE igw-eaad4883.

Documentation

createInternetGateway :: CreateInternetGateway

Minimum specification for a CreateInternetGateway request.

data CreateInternetGateway

Constructors

CreateInternetGateway

Fields

_cigrDryRun :: Maybe Bool

cigrDryRun :: Iso' CreateInternetGateway (Maybe Bool)

data CreateInternetGatewayResponse

Constructors

CreateInternetGatewayResponse

Fields

_cigsInternetGateway :: Maybe InternetGateway

Information about the Internet gateway.

cigsInternetGateway :: Iso' CreateInternetGatewayResponse (Maybe InternetGateway)