Network.AWS.EC2.V2014_06_15.CreateRouteTable

Description

Creates a route table for the specified VPC. After you create a route table, you can add routes and associate the table with a subnet. For more information about route tables, see Route Tables in the Amazon Virtual Private Cloud User Guide. Example This example creates a route table for the VPC with the ID vpc-11ad4878. By default, every route table includes a local route that enables traffic to flow within the VPC. The following response shows that route. https:/ec2.amazonaws.com?Action=CreateRouteTable &amp;VpcId=vpc-11ad4878 &amp;AUTHPARAMS 59dbff89-35bd-4eac-99ed-be587EXAMPLE rtb-f9ad4890 vpc-11ad4878 10.0.0.0/22 local active.

Documentation

createRouteTable

Arguments

:: Text
_crtrVpcId

-> CreateRouteTable

Minimum specification for a CreateRouteTable request.

data CreateRouteTable

Constructors

CreateRouteTable

Fields

_crtrVpcId :: Text

The ID of the VPC.

_crtrDryRun :: Maybe Bool

crtrVpcId :: Lens' CreateRouteTable Text

crtrDryRun :: Lens' CreateRouteTable (Maybe Bool)

data CreateRouteTableResponse

Constructors

CreateRouteTableResponse

Fields

_crtsRouteTable :: Maybe RouteTable

Information about the route table.

crtsRouteTable :: Iso' CreateRouteTableResponse (Maybe RouteTable)