Network.AWS.EC2.V2014_06_15.CreateSnapshot

Description

Creates a snapshot of an Amazon EBS volume and stores it in Amazon S3. You can use snapshots for backups, to make copies of Amazon EBS volumes, and to save data before shutting down an instance. When a snapshot is created, any AWS Marketplace product codes that are associated with the source volume are propagated to the snapshot. You can take a snapshot of an attached volume that is in use. However, snapshots only capture data that has been written to your Amazon EBS volume at the time the snapshot command is issued; this may exclude any data that has been cached by any applications or the operating system. If you can pause any file writes to the volume long enough to take a snapshot, your snapshot should be complete. However, if you cannot pause all file writes to the volume, you should unmount the volume from within the instance, issue the snapshot command, and then remount the volume to ensure a consistent and complete snapshot. You may remount and use your volume while the snapshot status is pending. To create a snapshot for Amazon EBS volumes that serve as root devices, you should stop the instance before taking the snapshot. Snapshots that are taken from encrypted volumes are automatically encrypted. Volumes that are created from encrypted snapshots are also automatically encrypted. Your encrypted volumes and any associated snapshots always remain protected. For more information, see Amazon Elastic Block Store and Amazon EBS Encryption in the Amazon Elastic Compute Cloud User Guide. Example This example creates a snapshot of the volume with the ID vol-1a2b3c4d. https:/ec2.amazonaws.com?Action=CreateSnapshot &amp;VolumeId=vol-1a2b3c4d &amp;Description=Daily+Backup &amp;AUTHPARAMS &lt;CreateSnapshotResponse xmlns="http://ec2.amazonaws.com/doc/2014-05-01/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;snapshotId&gt;snap-1a2b3c4d&lt;/snapshotId&gt; &lt;volumeId&gt;vol-1a2b3c4d&lt;/volumeId&gt; &lt;status&gt;pending&lt;/status&gt; &lt;startTime&gt;YYYY-MM-DDTHH:MM:SS.000Z&lt;/startTime&gt; &lt;progress&gt;60%&lt;/progress&gt; &lt;ownerId&gt;111122223333&lt;/ownerId&gt; &lt;volumeSize&gt;30&lt;/volumeSize&gt; &lt;description&gt;Daily Backup&lt;description&gt; &lt;CreateSnapshotResponse&gt;.

Documentation

createSnapshot

Arguments

:: Text
_csvVolumeId

-> CreateSnapshot

Minimum specification for a CreateSnapshot request.

data CreateSnapshot

Constructors

CreateSnapshot

Fields

_csvVolumeId :: Text

The ID of the Amazon EBS volume.

_csvDryRun :: Maybe Bool

_csvDescription :: Maybe Text

A description for the snapshot.

csvVolumeId :: Lens' CreateSnapshot Text

csvDryRun :: Lens' CreateSnapshot (Maybe Bool)

csvDescription :: Lens' CreateSnapshot (Maybe Text)

data CreateSnapshotResponse

Constructors

CreateSnapshotResponse

Fields

_ssyEncrypted :: Maybe Bool

Indicates whether the snapshot is encrypted.

_ssyStartTime :: Maybe ISO8601

The time stamp when the snapshot was initiated.

_ssyVolumeSize :: Maybe Integer

The size of the volume, in GiB.

_ssyState :: Maybe SnapshotState

The snapshot state.

_ssyOwnerAlias :: Maybe Text

The AWS account alias (for example, amazon, self) or AWS account ID that owns the snapshot.

_ssyProgress :: Maybe Text

The progress of the snapshot, as a percentage.

_ssyOwnerId :: Maybe Text

The AWS account ID of the Amazon EBS snapshot owner.

_ssyVolumeId :: Maybe Text

The ID of the volume.

_ssyDescription :: Maybe Text

The description for the snapshot.

_ssySnapshotId :: Maybe Text

The ID of the snapshot.

ssyVolumeSize :: Lens' CreateSnapshotResponse (Maybe Integer)

ssyVolumeId :: Lens' CreateSnapshotResponse (Maybe Text)

ssyState :: Lens' CreateSnapshotResponse (Maybe SnapshotState)

ssyStartTime :: Lens' CreateSnapshotResponse (Maybe ISO8601)

ssySnapshotId :: Lens' CreateSnapshotResponse (Maybe Text)

ssyProgress :: Lens' CreateSnapshotResponse (Maybe Text)

ssyOwnerId :: Lens' CreateSnapshotResponse (Maybe Text)

ssyOwnerAlias :: Lens' CreateSnapshotResponse (Maybe Text)

ssyEncrypted :: Lens' CreateSnapshotResponse (Maybe Bool)

ssyDescription :: Lens' CreateSnapshotResponse (Maybe Text)