Network.AWS.EC2.V2014_06_15.CreateVpc

Description

Creates a VPC with the specified CIDR block. The smallest VPC you can create uses a 28 netmask (16 IP addresses), and the largest uses a 16 netmask (65,536 IP addresses). To help you decide how big to make your VPC, see Your VPC and Subnets in the Amazon Virtual Private Cloud User Guide. By default, each instance you launch in the VPC has the default DHCP options, which includes only a default DNS server that we provide (AmazonProvidedDNS). For more information about DHCP options, see DHCP Options Sets in the Amazon Virtual Private Cloud User Guide. Example 1 This example creates a VPC with the CIDR block 10.0.0.0/16. https:/ec2.amazonaws.com?Action=CreateVpc &amp;CidrBlock=10.0.0.0/16 &amp;AUTHPARAMS &lt;CreateVpcResponse xmlns="http://ec2.amazonaws.com/doc/2013-10-01/"&gt; &lt;requestId&gt;7a62c49f-347e-4fc4-9331-6e8eEXAMPLE&lt;/requestId&gt; &lt;vpc&gt; &lt;vpcId&gt;vpc-1a2b3c4d&lt;/vpcId&gt; &lt;state&gt;pending&lt;/state&gt; &lt;cidrBlock&gt;10.0.0.016&lt;cidrBlock&gt; &lt;dhcpOptionsId&gt;dopt-1a2b3c4d2&lt;/dhcpOptionsId&gt; &lt;instanceTenancy&gt;default&lt;instanceTenancy&gt; &lt;tagSet&gt; &lt;vpc&gt; &lt;CreateVpcResponse&gt; Example 2 This example creates a VPC with the dedicated tenancy option. https:/ec2.amazonaws.com?Action=CreateVpc &amp;CidrBlock=10.32.0.0/16 &amp;InstanceTenancy=dedicated &amp;AUTHPARAMS &lt;CreateVpcResponse xmlns="http://ec2.amazonaws.com/doc/2013-10-01/"&gt; &lt;requestId&gt;a9e49797-a74f-4f68-b302-a134a51fd054&lt;/requestId&gt; &lt;vpc&gt; &lt;vpcId&gt;vpc-11a63c78&lt;/vpcId&gt; &lt;state&gt;pending&lt;/state&gt; &lt;cidrBlock&gt;10.32.0.016&lt;cidrBlock&gt; &lt;dhcpOptionsId&gt;dopt-1a2b3c4d2&lt;/dhcpOptionsId&gt; &lt;instanceTenancy&gt;dedicated&lt;instanceTenancy&gt; &lt;vpc&gt; &lt;/CreateVpcResponse&gt;.

Documentation

createVpc

Arguments

:: Text
_cvrCidrBlock

-> CreateVpc

Minimum specification for a CreateVpc request.

data CreateVpc

Constructors

CreateVpc

Fields

_cvrCidrBlock :: Text

The network range for the VPC, in CIDR notation. For example, 10.0.0.0/16.

_cvrDryRun :: Maybe Bool

_cvrInstanceTenancy :: Maybe Tenancy

The supported tenancy options for instances launched into the VPC. A value of default means that instances can be launched with any tenancy; a value of dedicated means all instances launched into the VPC are launched as dedicated tenancy instances regardless of the tenancy assigned to the instance at launch. Dedicated tenancy instances runs on single-tenant hardware. Default: default.

cvrInstanceTenancy :: Lens' CreateVpc (Maybe Tenancy)

cvrDryRun :: Lens' CreateVpc (Maybe Bool)

cvrCidrBlock :: Lens' CreateVpc Text

data CreateVpcResponse

Constructors

CreateVpcResponse

Fields

_cvsVpc :: Maybe Vpc

Information about the VPC.

cvsVpc :: Iso' CreateVpcResponse (Maybe Vpc)