Network.AWS.EC2.V2014_06_15.CreateVpnGateway

Description

Creates a virtual private gateway. A virtual private gateway is the endpoint on the VPC side of your VPN connection. You can create a virtual private gateway before creating the VPC itself. For more information about virtual private gateways, see Adding a Hardware Virtual Private Gateway to Your VPC in the Amazon Virtual Private Cloud User Guide. Example This example creates a virtual private gateway. https:/ec2.amazonaws.com?Action=CreateVpnGateway &amp;Type=ipsec.1 &amp;AUTHPARAMS &lt;CreateVpnGatewayResponse xmlns="http://ec2.amazonaws.com/doc/2013-10-01/"&gt; &lt;requestId&gt;7a62c49f-347e-4fc4-9331-6e8eEXAMPLE&lt;/requestId&gt; &lt;vpnGateway&gt; &lt;vpnGatewayId&gt;vgw-8db04f81&lt;/vpnGatewayId&gt; &lt;state&gt;pending&lt;state&gt; &lt;type&gt;ipsec.1&lt;type&gt; &lt;availabilityZone&gt;us-east-1a&lt;/availabilityZone&gt; &lt;attachments&gt; &lt;tagSet&gt; &lt;/vpnGateway&gt; &lt;/CreateVpnGatewayResponse&gt;.

Documentation

createVpnGateway

Arguments

:: GatewayType
_cvgrType

-> CreateVpnGateway

Minimum specification for a CreateVpnGateway request.

data CreateVpnGateway

Constructors

CreateVpnGateway

Fields

_cvgrType :: GatewayType

The type of VPN connection this virtual private gateway supports.

_cvgrDryRun :: Maybe Bool

_cvgrAvailabilityZone :: Maybe Text

The Availability Zone for the virtual private gateway.

cvgrType :: Lens' CreateVpnGateway GatewayType

cvgrDryRun :: Lens' CreateVpnGateway (Maybe Bool)

cvgrAvailabilityZone :: Lens' CreateVpnGateway (Maybe Text)

data CreateVpnGatewayResponse

Constructors

CreateVpnGatewayResponse

Fields

_cvgsVpnGateway :: Maybe VpnGateway

Information about the virtual private gateway.

cvgsVpnGateway :: Iso' CreateVpnGatewayResponse (Maybe VpnGateway)