Network.AWS.EC2.V2014_06_15.DeleteDhcpOptions

Description

Deletes the specified set of DHCP options. You must disassociate the set of DHCP options before you can delete it. You can disassociate the set of DHCP options by associating either a new set of options or the default set of options with the VPC. Example This example deletes the specified set of DHCP options. https:/ec2.amazonaws.com?Action=DeleteDhcpOptions &amp;DhcpOptionsId=dopt-7a8b9c2d &amp;AUTHPARAMS &lt;DeleteDhcpOptionsResponse xmlns="http://ec2.amazonaws.com/doc/2013-10-01/"&gt; &lt;requestId&gt;7a62c49f-347e-4fc4-9331-6e8eEXAMPLE&lt;/requestId&gt; &lt;return&gt;true&lt;return&gt; &lt;DeleteDhcpOptionsResponse&gt;.

Documentation

deleteDhcpOptions

Arguments

:: Text
_ddotDhcpOptionsId

-> DeleteDhcpOptions

Minimum specification for a DeleteDhcpOptions request.

data DeleteDhcpOptions

Constructors

DeleteDhcpOptions

Fields

_ddotDhcpOptionsId :: Text

The ID of the DHCP options set.

_ddotDryRun :: Maybe Bool

ddotDryRun :: Lens' DeleteDhcpOptions (Maybe Bool)

ddotDhcpOptionsId :: Lens' DeleteDhcpOptions Text

data DeleteDhcpOptionsResponse

Constructors

DeleteDhcpOptionsResponse