Network.AWS.EC2.V2014_06_15.DeleteKeyPair

Description

Deletes the specified key pair, by removing the public key from Amazon EC2. Example This example request deletes the key pair named my-key-pair. https:/ec2.amazonaws.com?Action=DeleteKeyPair &amp;KeyName=my-key-pair &amp;AUTHPARAMS &lt;DeleteKeyPairResponse xmlns="http://ec2.amazonaws.com/doc/2013-10-01/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;return&gt;true&lt;return&gt; &lt;DeleteKeyPairResponse&gt;.

Documentation

deleteKeyPair

Arguments

:: Text
_dkprKeyName

-> DeleteKeyPair

Minimum specification for a DeleteKeyPair request.

data DeleteKeyPair

Constructors

DeleteKeyPair

Fields

_dkprKeyName :: Text

The name of the key pair.

_dkprDryRun :: Maybe Bool

dkprKeyName :: Lens' DeleteKeyPair Text

dkprDryRun :: Lens' DeleteKeyPair (Maybe Bool)

data DeleteKeyPairResponse

Constructors

DeleteKeyPairResponse