Network.AWS.EC2.V2014_06_15.DeleteSnapshot

Description

Deletes the specified snapshot. When you make periodic snapshots of a volume, the snapshots are incremental, and only the blocks on the device that have changed since your last snapshot are saved in the new snapshot. When you delete a snapshot, only the data not needed for any other snapshot is removed. So regardless of which prior snapshots have been deleted, all active snapshots will have access to all the information needed to restore the volume. You cannot delete a snapshot of the root device of an Amazon EBS volume used by a registered AMI. You must first de-register the AMI before you can delete the snapshot. For more information, see Deleting an Amazon EBS Snapshot in the Amazon Elastic Compute Cloud User Guide. Example This example request deletes the snapshot with the ID snap-1a2b3c4d. https:/ec2.amazonaws.com?Action=DeleteSnapshot &amp;SnapshotId.1=snap-1a2b3c4d &amp;AUTHPARAMS &lt;DeleteSnapshotResponse xmlns="http://ec2.amazonaws.com/doc/2014-05-01/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;return&gt;true&lt;return&gt; &lt;DeleteSnapshotResponse&gt;.

Documentation

deleteSnapshot

Arguments

:: Text
_dsySnapshotId

-> DeleteSnapshot

Minimum specification for a DeleteSnapshot request.

data DeleteSnapshot

Constructors

DeleteSnapshot

Fields

_dsySnapshotId :: Text

The ID of the Amazon EBS snapshot.

_dsyDryRun :: Maybe Bool

dsySnapshotId :: Lens' DeleteSnapshot Text

dsyDryRun :: Lens' DeleteSnapshot (Maybe Bool)

data DeleteSnapshotResponse

Constructors

DeleteSnapshotResponse