Network.AWS.EC2.V2014_06_15.DeleteVpnConnectionRoute

Description

Deletes the specified static route associated with a VPN connection between an existing virtual private gateway and a VPN customer gateway. The static route allows traffic to be routed from the virtual private gateway to the VPN customer gateway. Example This example deletes a static route to the destination CIDR block 11.12.0.0/16 associated with the VPN connection with the ID vpn-83ad48ea. Note that when using the Query API, the "/" is denoted as "%2F". https:/ec2.amazonaws.com?Action=DeleteVpnConnectionRoute &amp;DestinationCidrBlock=11.12.0.0%2F16 &amp;VpnConnectionId=vpn-83ad48ea &amp;AUTHPARAMS &lt;DeleteVpnConnectionRouteResponse xmlns="http://ec2.amazonaws.com/doc/2013-10-01/"&gt; &lt;requestId&gt;4f35a1b2-c2c3-4093-b51f-abb9d7311990&lt;/requestId&gt; &lt;return&gt;true&lt;/return&gt; &lt;/DeleteVpnConnectionRouteResponse&gt;.

Documentation

data DeleteVpnConnectionRoute

Constructors

DeleteVpnConnectionRoute

Fields

_dvcrrVpnConnectionId :: Text

The ID of the VPN connection.

_dvcrrDestinationCidrBlock :: Text

The CIDR block associated with the local subnet of the customer network.

dvcrrVpnConnectionId :: Lens' DeleteVpnConnectionRoute Text

dvcrrDestinationCidrBlock :: Lens' DeleteVpnConnectionRoute Text

data DeleteVpnConnectionRouteResponse

Constructors

DeleteVpnConnectionRouteResponse