Network.AWS.EC2.V2014_06_15.DescribeAvailabilityZones

Description

Describes one or more of the Availability Zones that are available to you. The results include zones only for the region you're currently using. If there is an event impacting an Availability Zone, you can use this request to view the state and any provided message for that Availability Zone. For more information, see Regions and Availability Zones in the Amazon Elastic Compute Cloud User Guide. Example This example request describes the Availability Zones that are available to you. The response includes Availability Zones only for the current region. https:/ec2.amazonaws.com?Action=DescribeAvailabilityZones &amp;AUTHPARAMS &lt;DescribeAvailabilityZonesResponse xmlns="http://ec2.amazonaws.com/doc/2013-10-01/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;availabilityZoneInfo&gt; &lt;item&gt; &lt;zoneName&gt;us-east-1a&lt;/zoneName&gt; &lt;zoneState&gt;available&lt;/zoneState&gt; &lt;regionName&gt;us-east-1&lt;regionName&gt; &lt;messageSet&gt; &lt;item&gt; &lt;item&gt; &lt;zoneName&gt;us-east-1b&lt;zoneName&gt; &lt;zoneState&gt;available&lt;/zoneState&gt; &lt;regionName&gt;us-east-1&lt;regionName&gt; &lt;messageSet&gt; &lt;item&gt; &lt;item&gt; &lt;zoneName&gt;us-east-1c&lt;zoneName&gt; &lt;zoneState&gt;available&lt;/zoneState&gt; &lt;regionName&gt;us-east-1&lt;regionName&gt; &lt;messageSet&gt; &lt;item&gt; &lt;item&gt; &lt;zoneName&gt;us-east-1d&lt;zoneName&gt; &lt;zoneState&gt;available&lt;/zoneState&gt; &lt;regionName&gt;us-east-1&lt;regionName&gt; &lt;messageSet&gt; &lt;item&gt; &lt;availabilityZoneInfo&gt; &lt;/DescribeAvailabilityZonesResponse&gt;.

Documentation

describeAvailabilityZones :: DescribeAvailabilityZones

Minimum specification for a DescribeAvailabilityZones request.

data DescribeAvailabilityZones

Constructors

DescribeAvailabilityZones

Fields

_dazrDryRun :: Maybe Bool

_dazrFilters :: [Filter]

One or more filters. message - Information about the Availability Zone. region-name - The name of the region for the Availability Zone (for example, us-east-1). state - The state of the Availability Zone (available | impaired | unavailable). zone-name - The name of the Availability Zone (for example, us-east-1a).

_dazrZoneNames :: [Text]

The names of one or more Availability Zones.

dazrZoneNames :: Lens' DescribeAvailabilityZones [Text]

dazrFilters :: Lens' DescribeAvailabilityZones [Filter]

dazrDryRun :: Lens' DescribeAvailabilityZones (Maybe Bool)

data DescribeAvailabilityZonesResponse

Constructors

DescribeAvailabilityZonesResponse

Fields

_dazsAvailabilityZones :: [AvailabilityZone]

Information about one or more Availability Zones.

dazsAvailabilityZones :: Iso' DescribeAvailabilityZonesResponse [AvailabilityZone]