Network.AWS.EC2.V2014_06_15.DescribeImageAttribute

Description

Describes the specified attribute of the specified AMI. You can specify only one attribute at a time. Example 1 This example lists the launch permissions for the specified AMI. https:/ec2.amazonaws.com?Action=DescribeImageAttribute &amp;ImageId=ami-61a54008 &amp;Attribute=launchPermission &amp;AUTHPARAMS 59dbff89-35bd-4eac-99ed-be587EXAMPLE ami-61a54008 all 495219933132 Example 2 This example lists the product codes for the specified AMI. https:/ec2.amazonaws.com?Action=DescribeImageAttribute &amp;ImageId=ami-2bb65342 &amp;Attribute=productCodes &amp;AUTHPARAMS 59dbff89-35bd-4eac-99ed-be587EXAMPLE ami-2bb65342 a1b2c3d4e5f6g7h8i9j10k11 marketplace.

Documentation

describeImageAttribute

Arguments

:: ImageAttributeName
_diasAttribute

-> Text
_diasImageId

-> DescribeImageAttribute

Minimum specification for a DescribeImageAttribute request.

data DescribeImageAttribute

Constructors

DescribeImageAttribute

Fields

_diasAttribute :: ImageAttributeName

The AMI attribute.

_diasImageId :: Text

The ID of the AMI.

_diasDryRun :: Maybe Bool

diasImageId :: Lens' DescribeImageAttribute Text

diasDryRun :: Lens' DescribeImageAttribute (Maybe Bool)

diasAttribute :: Lens' DescribeImageAttribute ImageAttributeName

data DescribeImageAttributeResponse

Constructors

DescribeImageAttributeResponse

Fields

_ibRamdiskId :: Maybe AttributeValue

The RAM disk ID.

_ibKernelId :: Maybe AttributeValue

The kernel ID.

_ibSriovNetSupport :: Maybe AttributeValue

_ibDescription :: Maybe AttributeValue

A description for the AMI.

_ibBlockDeviceMappings :: [BlockDeviceMapping]

One or more block device mapping entries.

_ibLaunchPermissions :: [LaunchPermission]

One or more launch permissions.

_ibProductCodes :: [ProductCode]

One or more product codes.

_ibImageId :: Maybe Text

The ID of the AMI.

ibSriovNetSupport :: Lens' DescribeImageAttributeResponse (Maybe AttributeValue)

ibRamdiskId :: Lens' DescribeImageAttributeResponse (Maybe AttributeValue)

ibProductCodes :: Lens' DescribeImageAttributeResponse [ProductCode]

ibLaunchPermissions :: Lens' DescribeImageAttributeResponse [LaunchPermission]

ibKernelId :: Lens' DescribeImageAttributeResponse (Maybe AttributeValue)

ibImageId :: Lens' DescribeImageAttributeResponse (Maybe Text)

ibDescription :: Lens' DescribeImageAttributeResponse (Maybe AttributeValue)

ibBlockDeviceMappings :: Lens' DescribeImageAttributeResponse [BlockDeviceMapping]