Network.AWS.EC2.V2014_06_15.DescribeInternetGateways

Description

Describes one or more of your Internet gateways. Example This example describes all your Internet gateways. https:/ec2.amazonaws.com?Action=DescribeInternetGateways &amp;AUTHPARAMS 59dbff89-35bd-4eac-99ed-be587EXAMPLE igw-eaad4883EXAMPLE vpc-11ad4878 available.

Documentation

describeInternetGateways :: DescribeInternetGateways

Minimum specification for a DescribeInternetGateways request.

data DescribeInternetGateways

Constructors

DescribeInternetGateways

Fields

_digsDryRun :: Maybe Bool

_digsFilters :: [Filter]

One or more filters. attachment.state - The current state of the attachment between the gateway and the VPC. Present only if a VPC is attached. attachment.vpc-id - The ID of an attached VPC. internet-gateway-id - The ID of the Internet gateway. tag:key=value - The key/value combination of a tag assigned to the resource. tag-key - The key of a tag assigned to the resource. This filter is independent of the tag-value filter. For example, if you use both the filter "tag-key=Purpose" and the filter "tag-value=X", you get any resources assigned both the tag key Purpose (regardless of what the tag's value is), and the tag value X (regardless of what the tag's key is). If you want to list only resources where Purpose is X, see the tag:key=value filter. tag-value - The value of a tag assigned to the resource. This filter is independent of the tag-key filter.

_digsInternetGatewayIds :: [Text]

One or more Internet gateway IDs. Default: Describes all your Internet gateways.

digsInternetGatewayIds :: Lens' DescribeInternetGateways [Text]

digsFilters :: Lens' DescribeInternetGateways [Filter]

digsDryRun :: Lens' DescribeInternetGateways (Maybe Bool)

data DescribeInternetGatewaysResponse

Constructors

DescribeInternetGatewaysResponse

Fields

_digtInternetGateways :: [InternetGateway]

Information about one or more Internet gateways.

digtInternetGateways :: Iso' DescribeInternetGatewaysResponse [InternetGateway]