Network.AWS.EC2.V2014_06_15.DescribePlacementGroups

Description

Describes one or more of your placement groups. For more information about placement groups and cluster instances, see Cluster Instances in the Amazon Elastic Compute Cloud User Guide. Example This example describes the placement group named XYZ-cluster. https:/ec2.amazonaws.com?Action=DescribePlacementGroups &amp;GroupName.1=XYZ-cluster &amp;AUTHPARAMS d4904fd9-82c2-4ea5-adfe-a9cc3EXAMPLE XYZ-cluster cluster available Example This example filters the response to include only placement groups that include the string Project in the name. https:/ec2.amazonaws.com?Action=DescribePlacementGroups &amp;Filter.1.Name=group-name &amp;Filter.1.Value=*Project* &amp;AUTHPARAMS d4904fd9-82c2-4ea5-adfe-a9cc3EXAMPLE Project-cluster cluster available.

Documentation

describePlacementGroups :: DescribePlacementGroups

Minimum specification for a DescribePlacementGroups request.

data DescribePlacementGroups

Constructors

DescribePlacementGroups

Fields

_dpgsDryRun :: Maybe Bool

_dpgsFilters :: [Filter]

One or more filters. group-name - The name of the placement group. state - The state of the placement group (pending | available | deleting | deleted). strategy - The strategy of the placement group (cluster).

_dpgsGroupNames :: [Text]

One or more placement group names. Default: Describes all your placement groups, or only those otherwise specified.

dpgsGroupNames :: Lens' DescribePlacementGroups [Text]

dpgsFilters :: Lens' DescribePlacementGroups [Filter]

dpgsDryRun :: Lens' DescribePlacementGroups (Maybe Bool)

data DescribePlacementGroupsResponse

Constructors

DescribePlacementGroupsResponse

Fields

_dpgtPlacementGroups :: [PlacementGroup]

One or more placement groups.

dpgtPlacementGroups :: Iso' DescribePlacementGroupsResponse [PlacementGroup]