Network.AWS.EC2.V2014_06_15.DescribeRegions

Description

Describes one or more regions that are currently available to you. For a list of the regions supported by Amazon EC2, see Regions and Endpoints. Example 1 This example displays information about all regions. https:/ec2.amazonaws.com?Action=DescribeRegions &amp;AUTHPARAMS Example 2 This example displays information about the specified regions only. https:/ec2.amazonaws.com?Action=DescribeRegions &amp;RegionName.1=us-east-1 &amp;RegionName.2=eu-west-1 &amp;AUTHPARAMS 59dbff89-35bd-4eac-99ed-be587EXAMPLE us-east-1 ec2.us-east-1.amazonaws.com eu-west-1 ec2.eu-west-1amazonaws.com.

Documentation

describeRegions :: DescribeRegions

Minimum specification for a DescribeRegions request.

data DescribeRegions

Constructors

DescribeRegions

Fields

_drsDryRun :: Maybe Bool

_drsFilters :: [Filter]

One or more filters. endpoint - The endpoint of the region (for example, ec2.us-east-1.amazonaws.com). region-name - The name of the region (for example, us-east-1).

_drsRegionNames :: [Text]

The names of one or more regions.

drsRegionNames :: Lens' DescribeRegions [Text]

drsFilters :: Lens' DescribeRegions [Filter]

drsDryRun :: Lens' DescribeRegions (Maybe Bool)

data DescribeRegionsResponse

Constructors

DescribeRegionsResponse

Fields

_drtRegions :: [Region]

Information about one or more regions.

drtRegions :: Iso' DescribeRegionsResponse [Region]