Network.AWS.EC2.V2014_06_15.DescribeSnapshotAttribute

Description

Describes the specified attribute of the specified snapshot. You can specify only one attribute at a time. For more information about Amazon EBS snapshots, see Amazon EBS Snapshots in the Amazon Elastic Compute Cloud User Guide. Example This example describes permissions for a snapshot with the ID of snap-1a2b3c4d. https:/ec2.amazonaws.com?Action=DescribeSnapshotAttribute &amp;SnapshotId=snap-1a2b3c4d &amp;Attribute=createVolumePermission &amp;AUTHPARAMS 59dbff89-35bd-4eac-99ed-be587EXAMPLE snap-1a2b3c4d all.

Documentation

describeSnapshotAttribute

Arguments

:: SnapshotAttributeName
_dsarAttribute

-> Text
_dsarSnapshotId

-> DescribeSnapshotAttribute

Minimum specification for a DescribeSnapshotAttribute request.

data DescribeSnapshotAttribute

Constructors

DescribeSnapshotAttribute

Fields

_dsarAttribute :: SnapshotAttributeName

The snapshot attribute you would like to view.

_dsarSnapshotId :: Text

The ID of the Amazon EBS snapshot.

_dsarDryRun :: Maybe Bool

dsarSnapshotId :: Lens' DescribeSnapshotAttribute Text

dsarDryRun :: Lens' DescribeSnapshotAttribute (Maybe Bool)

dsarAttribute :: Lens' DescribeSnapshotAttribute SnapshotAttributeName

data DescribeSnapshotAttributeResponse

Constructors

DescribeSnapshotAttributeResponse

Fields

_dsasCreateVolumePermissions :: [CreateVolumePermission]

A list of permissions for creating volumes from the snapshot.

_dsasProductCodes :: [ProductCode]

A list of product codes.

_dsasSnapshotId :: Maybe Text

The ID of the Amazon EBS snapshot.

dsasSnapshotId :: Lens' DescribeSnapshotAttributeResponse (Maybe Text)

dsasProductCodes :: Lens' DescribeSnapshotAttributeResponse [ProductCode]

dsasCreateVolumePermissions :: Lens' DescribeSnapshotAttributeResponse [CreateVolumePermission]