Network.AWS.EC2.V2014_06_15.DescribeSpotDatafeedSubscription

Description

Describes the datafeed for Spot Instances. For more information, see Spot Instances in the Amazon Elastic Compute Cloud User Guide. Example This example describes the datafeed for the account. https:/ec2.amazonaws.com?Action=DescribeSpotDatafeedSubscription &amp;AUTHPARAMS 59dbff89-35bd-4eac-99ed-be587EXAMPLE 123456789012 my-s3-bucket spotdata_ Active.

Documentation

describeSpotDatafeedSubscription :: DescribeSpotDatafeedSubscription

Minimum specification for a DescribeSpotDatafeedSubscription request.

data DescribeSpotDatafeedSubscription

Constructors

DescribeSpotDatafeedSubscription

Fields

_dsdsrDryRun :: Maybe Bool

dsdsrDryRun :: Iso' DescribeSpotDatafeedSubscription (Maybe Bool)

data DescribeSpotDatafeedSubscriptionResponse

Constructors

DescribeSpotDatafeedSubscriptionResponse

Fields

_dsdssSpotDatafeedSubscription :: Maybe SpotDatafeedSubscription

The Spot Instance datafeed subscription.

dsdssSpotDatafeedSubscription :: Iso' DescribeSpotDatafeedSubscriptionResponse (Maybe SpotDatafeedSubscription)