Network.AWS.EC2.V2014_06_15.DetachVpnGateway

Description

Detaches a virtual private gateway from a VPC. You do this if you're planning to turn off the VPC and not use it anymore. You can confirm a virtual private gateway has been completely detached from a VPC by describing the virtual private gateway (any attachments to the virtual private gateway are also described). You must wait for the attachment's state to switch to detached before you can delete the VPC or attach a different VPC to the virtual private gateway. Example This example detaches the specified virtual private gateway from the specified VPC. https:/ec2.amazonaws.com?Action=DetachVpnGateway &amp;VpnGatewayId=vgw-8db04f81 &amp;VpcId=vpc-1a2b3c4d &amp;AUTHPARAMS &lt;DetachVpnGatewayResponse xmlns="http://ec2.amazonaws.com/doc/2013-10-01/"&gt; &lt;requestId&gt;7a62c49f-347e-4fc4-9331-6e8eEXAMPLE&lt;/requestId&gt; &lt;return&gt;true&lt;return&gt; &lt;DetachVpnGatewayResponse&gt;.

Documentation

detachVpnGateway

Arguments

:: Text
_dvgsVpnGatewayId

-> Text
_dvgsVpcId

-> DetachVpnGateway

Minimum specification for a DetachVpnGateway request.

data DetachVpnGateway

Constructors

DetachVpnGateway

Fields

_dvgsVpnGatewayId :: Text

The ID of the virtual private gateway.

_dvgsVpcId :: Text

The ID of the VPC.

_dvgsDryRun :: Maybe Bool

dvgsVpnGatewayId :: Lens' DetachVpnGateway Text

dvgsVpcId :: Lens' DetachVpnGateway Text

dvgsDryRun :: Lens' DetachVpnGateway (Maybe Bool)

data DetachVpnGatewayResponse

Constructors

DetachVpnGatewayResponse