Network.AWS.EC2.V2014_06_15.DisableVgwRoutePropagation

Description

Disables a virtual private gateway (VGW) from propagating routes to the routing tables of a VPC. Example This example disables the virtual private gateway vgw-d8e09e8a from automatically propagating routes to the routing table with ID rtb-c98a35a0. https:/ec2.amazonaws.com?Action=DisableVgwRoutePropagationResponse &amp;RouteTableID=rtb-c98a35a0 &amp;GatewayId= vgw-d8e09e8a &amp;AUTHPARAMS &lt;DisableVgwRoutePropagationResponse xmlns="http://ec2.amazonaws.com/doc/2013-10-01/"&gt; &lt;requestId&gt;4f35a1b2-c2c3-4093-b51f-abb9d7311990&lt;/requestId&gt; &lt;return&gt;true&lt;/return&gt; &lt;/DisableVgwRoutePropagationResponse&gt;.

Documentation

data DisableVgwRoutePropagation

Constructors

DisableVgwRoutePropagation

Fields

_dvrprRouteTableId :: Text

The ID of the routing table.

_dvrprGatewayId :: Text

The ID of the virtual private gateway.

dvrprRouteTableId :: Lens' DisableVgwRoutePropagation Text

dvrprGatewayId :: Lens' DisableVgwRoutePropagation Text

data DisableVgwRoutePropagationResponse

Constructors

DisableVgwRoutePropagationResponse