Network.AWS.EC2.V2014_06_15.ModifyImageAttribute

Description

Modifies the specified attribute of the specified AMI. You can specify only one attribute at a time. AWS Marketplace product codes cannot be modified. Images with an AWS Marketplace product code cannot be made public. Example.

Documentation

modifyImageAttribute

Arguments

:: Text
_miarImageId

-> ModifyImageAttribute

Minimum specification for a ModifyImageAttribute request.

data ModifyImageAttribute

Constructors

ModifyImageAttribute

Fields

_miarImageId :: Text

The ID of the AMI.

_miarDescription :: Maybe AttributeValue

A description for the AMI.

_miarDryRun :: Maybe Bool

_miarLaunchPermission :: Maybe LaunchPermissionModifications

_miarProductCodes :: [Text]

One or more product codes. After you add a product code to an AMI, it can't be removed. This is only valid when modifying the productCodes attribute.

_miarUserGroups :: [Text]

One or more user groups. This is only valid when modifying the launchPermission attribute.

_miarUserIds :: [Text]

One or more AWS account IDs. This is only valid when modifying the launchPermission attribute.

_miarAttribute :: Maybe Text

The name of the attribute to modify.

_miarValue :: Maybe Text

The value of the attribute being modified. This is only valid when modifying the description attribute.

_miarOperationType :: Maybe Text

The operation type.

miarValue :: Lens' ModifyImageAttribute (Maybe Text)

miarUserIds :: Lens' ModifyImageAttribute [Text]

miarUserGroups :: Lens' ModifyImageAttribute [Text]

miarProductCodes :: Lens' ModifyImageAttribute [Text]

miarOperationType :: Lens' ModifyImageAttribute (Maybe Text)

miarLaunchPermission :: Lens' ModifyImageAttribute (Maybe LaunchPermissionModifications)

miarImageId :: Lens' ModifyImageAttribute Text

miarDryRun :: Lens' ModifyImageAttribute (Maybe Bool)

miarDescription :: Lens' ModifyImageAttribute (Maybe AttributeValue)

miarAttribute :: Lens' ModifyImageAttribute (Maybe Text)

data ModifyImageAttributeResponse

Constructors

ModifyImageAttributeResponse