Network.AWS.EC2.V2014_06_15.ReleaseAddress

Description

Releases the specified Elastic IP address. After releasing an Elastic IP address, it is released to the IP address pool and might be unavailable to you. Be sure to update your DNS records and any servers or devices that communicate with the address. If you attempt to release an Elastic IP address that you already released, you'll get an AuthFailure error if the address is already allocated to another AWS account. [EC2-Classic, default VPC] Releasing an Elastic IP address automatically disassociates it from any instance that it's associated with. To disassociate an Elastic IP address without releasing it, use DisassociateAddress. [Nondefault VPC] You must use the DisassociateAddress to disassociate the Elastic IP address before you try to release it. Otherwise, Amazon EC2 returns an error (InvalidIPAddress.InUse). Example for EC2-Classic This example releases the specified Elastic IP address for EC2-Classic. https:/ec2.amazonaws.com?Action=ReleaseAddress &amp;PublicIp=192.0.2.1 &amp;AUTHPARAMS Example for EC2-VPC This example releases the specified Elastic IP address for EC2-VPC. https:/ec2.amazonaws.com?Action=ReleaseAddress &amp;AllocationId=eipalloc-5723d13e &amp;AUTHPARAMS.

Documentation

releaseAddress :: ReleaseAddress

Minimum specification for a ReleaseAddress request.

data ReleaseAddress

Constructors

ReleaseAddress

Fields

_rarDryRun :: Maybe Bool

_rarAllocationId :: Maybe Text

EC2-VPC

The allocation ID.

_rarPublicIp :: Maybe Text

EC2-Classic

The Elastic IP address.

rarPublicIp :: Lens' ReleaseAddress (Maybe Text)

rarDryRun :: Lens' ReleaseAddress (Maybe Bool)

rarAllocationId :: Lens' ReleaseAddress (Maybe Text)

data ReleaseAddressResponse

Constructors

ReleaseAddressResponse