Network.AWS.EC2.V2014_06_15.ReplaceNetworkAclAssociation

Description

Changes which network ACL a subnet is associated with. By default when you create a subnet, it's automatically associated with the default network ACL. For more information about network ACLs, see Network ACLs in the Amazon Virtual Private Cloud User Guide. Example This example starts with a network ACL associated with a subnet, and a corresponding association ID aclassoc-e5b95c8c. You want to associate a different network ACL (acl-5fb85d36) with the subnet. The result is a new association ID representing the new association. https:/ec2.amazonaws.com?Action=ReplaceNetworkAclAssociation &amp;AssociationId=aclassoc-e5b95c8c &amp;NetworkAclId=acl-5fb85d36 &amp;AUTHPARAMS &lt;ReplaceNetworkAclAssociationResponse xmlns="http://ec2.amazonaws.com/doc/2013-10-01/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;newAssociationId&gt;aclassoc-17b85d7e&lt;/newAssociationId&gt; &lt;/ReplaceNetworkAclAssociationResponse&gt;.

Documentation

replaceNetworkAclAssociation

Arguments

:: Text
_rnaarAssociationId

-> Text
_rnaarNetworkAclId

-> ReplaceNetworkAclAssociation

Minimum specification for a ReplaceNetworkAclAssociation request.

data ReplaceNetworkAclAssociation

Constructors

ReplaceNetworkAclAssociation

Fields

_rnaarAssociationId :: Text

The ID of the current association between the original network ACL and the subnet.

_rnaarNetworkAclId :: Text

The ID of the new ACL to associate with the subnet.

_rnaarDryRun :: Maybe Bool

rnaarNetworkAclId :: Lens' ReplaceNetworkAclAssociation Text

rnaarDryRun :: Lens' ReplaceNetworkAclAssociation (Maybe Bool)

rnaarAssociationId :: Lens' ReplaceNetworkAclAssociation Text

data ReplaceNetworkAclAssociationResponse

Constructors

ReplaceNetworkAclAssociationResponse

Fields

_rnaasNewAssociationId :: Maybe Text

The ID of the new association.

rnaasNewAssociationId :: Iso' ReplaceNetworkAclAssociationResponse (Maybe Text)