Network.AWS.EC2.V2014_06_15.ReplaceRoute

Description

Replaces an existing route within a route table in a VPC. You must provide only one of the following: Internet gateway, NAT instance, VPC peering connection, or network interface. For more information about route tables, see Route Tables in the Amazon Virtual Private Cloud User Guide. Example This example replaces a route in the specified route table. The new route matches the CIDR 10.0.0.0/8 and sends the traffic to the virtual private gateway with the ID vgw-1d00376e. https:/ec2.amazonaws.com?Action=ReplaceRoute &amp;RouteTableId=rtb-e4ad488d &amp;DestinationCidrBlock=10.0.0.0/8 &amp;GatewayId=vgw-1d00376e &amp;AUTHPARAMS &lt;ReplaceRouteResponse xmlns="http://ec2.amazonaws.com/doc/2013-10-01/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;return&gt;true&lt;return&gt; &lt;ReplaceRouteResponse&gt;.

Documentation

replaceRoute

Arguments

:: Text
_rrrRouteTableId

-> Text
_rrrDestinationCidrBlock

-> ReplaceRoute

Minimum specification for a ReplaceRoute request.

data ReplaceRoute

Constructors

ReplaceRoute

Fields

_rrrRouteTableId :: Text

The ID of the route table.

_rrrDestinationCidrBlock :: Text

The CIDR address block used for the destination match. The value you provide must match the CIDR of an existing route in the table.

_rrrDryRun :: Maybe Bool

_rrrVpcPeeringConnectionId :: Maybe Text

The ID of a VPC peering connection.

_rrrInstanceId :: Maybe Text

The ID of a NAT instance in your VPC.

_rrrNetworkInterfaceId :: Maybe Text

The ID of a network interface.

_rrrGatewayId :: Maybe Text

The ID of an Internet gateway attached to your VPC.

rrrVpcPeeringConnectionId :: Lens' ReplaceRoute (Maybe Text)

rrrRouteTableId :: Lens' ReplaceRoute Text

rrrNetworkInterfaceId :: Lens' ReplaceRoute (Maybe Text)

rrrInstanceId :: Lens' ReplaceRoute (Maybe Text)

rrrGatewayId :: Lens' ReplaceRoute (Maybe Text)

rrrDryRun :: Lens' ReplaceRoute (Maybe Bool)

rrrDestinationCidrBlock :: Lens' ReplaceRoute Text

data ReplaceRouteResponse

Constructors

ReplaceRouteResponse