Network.AWS.EC2.V2014_06_15.ReportInstanceStatus

Description

Submits feedback about the status of an instance. The instance must be in the running state. If your experience with the instance differs from the instance status returned by DescribeInstanceStatus, use ReportInstanceStatus to report your experience with the instance. Amazon EC2 collects this information to improve the accuracy of status checks. Example 1 This example reports instance health state for two instances. https:/ec2.amazonaws.com?Action=ReportInstanceStatus &amp;Status=impaired &amp;InstanceId.0=i-9440effb &amp;InstanceId.1=i-0cf27c63 &amp;AUTHPARAMS Example 2 This example reports instance health state for two instances with reason codes. https:/ec2.amazonaws.com?Action=ReportInstanceStatus &amp;Description=Description+of+my+issue. &amp;Status=impaired &amp;InstanceId.0=i-9440effb &amp;InstanceId.1=i-0cf27c63 &amp;ReasonCode.0=instance-performance-network &amp;ReasonCode.1=instance-performance-disk &amp;AUTHPARAMS.

Documentation

reportInstanceStatus :: ReportInstanceStatus

Minimum specification for a ReportInstanceStatus request.

data ReportInstanceStatus

Constructors

ReportInstanceStatus

Fields

_risrDryRun :: Maybe Bool

_risrStartTime :: Maybe ISO8601

The time at which the reported instance health state began.

_risrEndTime :: Maybe ISO8601

The time at which the reported instance health state ended.

_risrInstances :: [Text]

One or more instances.

_risrReasonCodes :: [ReportInstanceReasonCodes]

One or more reason codes that describes the health state of your instance. instance-stuck-in-state: My instance is stuck in a state. unresponsive: My instance is unresponsive. not-accepting-credentials: My instance is not accepting my credentials. password-not-available: A password is not available for my instance. performance-network: My instance is experiencing performance problems which I believe are network related. performance-instance-store: My instance is experiencing performance problems which I believe are related to the instance stores. performance-ebs-volume: My instance is experiencing performance problems which I believe are related to an EBS volume. performance-other: My instance is experiencing performance problems. other: [explain using the description parameter].

_risrStatus :: Maybe ReportStatusType

The status of all instances listed.

_risrDescription :: Maybe Text

Descriptive text about the health state of your instance.

risrStatus :: Lens' ReportInstanceStatus (Maybe ReportStatusType)

risrStartTime :: Lens' ReportInstanceStatus (Maybe ISO8601)

risrReasonCodes :: Lens' ReportInstanceStatus [ReportInstanceReasonCodes]

risrInstances :: Lens' ReportInstanceStatus [Text]

risrEndTime :: Lens' ReportInstanceStatus (Maybe ISO8601)

risrDryRun :: Lens' ReportInstanceStatus (Maybe Bool)

risrDescription :: Lens' ReportInstanceStatus (Maybe Text)

data ReportInstanceStatusResponse

Constructors

ReportInstanceStatusResponse