Network.AWS.EC2.V2014_06_15.UnmonitorInstances

Description

Disables monitoring for a running instance. For more information about monitoring instances, see Monitoring Your Instances and Volumes in the Amazon Elastic Compute Cloud User Guide. Example This example disables monitoring for the specified instances. https:/ec2.amazonaws.com?Action=UnmonitorInstances &amp;InstanceId.1=i-43a4412a &amp;InstanceId.2=i-23a3397d &amp;AUTHPARAMS &lt;UnmonitorInstancesResponse xmlns="http://ec2.amazonaws.com/doc/2013-10-01/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;instancesSet&gt; &lt;item&gt; &lt;instanceId&gt;i-43a4412a&lt;/instanceId&gt; &lt;monitoring&gt; &lt;state&gt;disabled&lt;state&gt; &lt;monitoring&gt; &lt;/item&gt; &lt;item&gt; &lt;instanceId&gt;i-23a3397d&lt;/instanceId&gt; &lt;monitoring&gt; &lt;state&gt;disabled&lt;state&gt; &lt;monitoring&gt; &lt;item&gt; &lt;instancesSet&gt; &lt;/UnmonitorInstancesResponse&gt;.

Documentation

unmonitorInstances

Arguments

:: [Text]
_uirInstanceIds

-> UnmonitorInstances

Minimum specification for a UnmonitorInstances request.

data UnmonitorInstances

Constructors

UnmonitorInstances

Fields

_uirInstanceIds :: [Text]

One or more instance IDs.

_uirDryRun :: Maybe Bool

uirInstanceIds :: Lens' UnmonitorInstances [Text]

uirDryRun :: Lens' UnmonitorInstances (Maybe Bool)

data UnmonitorInstancesResponse

Constructors

UnmonitorInstancesResponse

Fields

_uisInstanceMonitorings :: [InstanceMonitoring]

Monitoring information for one or more instances.

uisInstanceMonitorings :: Iso' UnmonitorInstancesResponse [InstanceMonitoring]