Network.AWS.EC2InstanceConnect

Description

Derived from API version 2018-04-02 of the AWS service descriptions, licensed under Apache 2.0.

Amazon EC2 Instance Connect enables system administrators to publish one-time use SSH public keys to EC2, providing users a simple and secure way to connect to their instances.

Service Configuration

defaultService :: Service

API version 2018-04-02 of the Amazon EC2 Instance Connect SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by EC2InstanceConnect.

SerialConsoleSessionUnavailableException

_SerialConsoleSessionUnavailableException :: AsError a => Getting (First ServiceError) a ServiceError

Unable to start a serial console session. Please try again.

AuthException

_AuthException :: AsError a => Getting (First ServiceError) a ServiceError

Either your AWS credentials are not valid or you do not have access to the EC2 instance.

SerialConsoleSessionLimitExceededException

_SerialConsoleSessionLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The instance currently has 1 active serial console session. Only 1 session is supported at a time.

InvalidArgsException

_InvalidArgsException :: AsError a => Getting (First ServiceError) a ServiceError

One of the parameters is not valid.

SerialConsoleAccessDisabledException

_SerialConsoleAccessDisabledException :: AsError a => Getting (First ServiceError) a ServiceError

Your account is not authorized to use the EC2 Serial Console. To authorize your account, run the EnableSerialConsoleAccess API. For more information, see EnableSerialConsoleAccess in the Amazon EC2 API Reference.

ThrottlingException

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The requests were made too frequently and have been throttled. Wait a while and try again. To increase the limit on your request frequency, contact AWS Support.

ServiceException

_ServiceException :: AsError a => Getting (First ServiceError) a ServiceError

The service encountered an error. Follow the instructions in the error message and try again.

EC2InstanceTypeInvalidException

_EC2InstanceTypeInvalidException :: AsError a => Getting (First ServiceError) a ServiceError

The instance type is not supported for connecting via the serial console. Only Nitro instance types are currently supported.

EC2InstanceNotFoundException

_EC2InstanceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified instance was not found.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

SendSSHPublicKey

data SendSSHPublicKey

See: newSendSSHPublicKey smart constructor.

Constructors

SendSSHPublicKey' Text Text Text Text

newSendSSHPublicKey

Arguments

:: Text
$sel:instanceId:SendSSHPublicKey'

-> Text
$sel:instanceOSUser:SendSSHPublicKey'

-> Text
$sel:sSHPublicKey:SendSSHPublicKey'

-> Text
$sel:availabilityZone:SendSSHPublicKey'

-> SendSSHPublicKey

Create a value of SendSSHPublicKey with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceId:SendSSHPublicKey', sendSSHPublicKey_instanceId - The ID of the EC2 instance.

$sel:instanceOSUser:SendSSHPublicKey', sendSSHPublicKey_instanceOSUser - The OS user on the EC2 instance for whom the key can be used to authenticate.

$sel:sSHPublicKey:SendSSHPublicKey', sendSSHPublicKey_sSHPublicKey - The public key material. To use the public key, you must have the matching private key.

$sel:availabilityZone:SendSSHPublicKey', sendSSHPublicKey_availabilityZone - The Availability Zone in which the EC2 instance was launched.

data SendSSHPublicKeyResponse

See: newSendSSHPublicKeyResponse smart constructor.

Constructors

SendSSHPublicKeyResponse' (Maybe Text) (Maybe Bool) Int

newSendSSHPublicKeyResponse

Arguments

:: Int
$sel:httpStatus:SendSSHPublicKeyResponse'

-> SendSSHPublicKeyResponse

Create a value of SendSSHPublicKeyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:SendSSHPublicKeyResponse', sendSSHPublicKeyResponse_requestId - The ID of the request. Please provide this ID when contacting AWS Support for assistance.

$sel:success:SendSSHPublicKeyResponse', sendSSHPublicKeyResponse_success - Is true if the request succeeds and an error otherwise.

$sel:httpStatus:SendSSHPublicKeyResponse', sendSSHPublicKeyResponse_httpStatus - The response's http status code.

SendSerialConsoleSSHPublicKey

data SendSerialConsoleSSHPublicKey

See: newSendSerialConsoleSSHPublicKey smart constructor.

Constructors

SendSerialConsoleSSHPublicKey' (Maybe Natural) Text Text

newSendSerialConsoleSSHPublicKey

Arguments

:: Text
$sel:instanceId:SendSerialConsoleSSHPublicKey'

-> Text
$sel:sSHPublicKey:SendSerialConsoleSSHPublicKey'

-> SendSerialConsoleSSHPublicKey

Create a value of SendSerialConsoleSSHPublicKey with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serialPort:SendSerialConsoleSSHPublicKey', sendSerialConsoleSSHPublicKey_serialPort - The serial port of the EC2 instance. Currently only port 0 is supported.

Default: 0

$sel:instanceId:SendSerialConsoleSSHPublicKey', sendSerialConsoleSSHPublicKey_instanceId - The ID of the EC2 instance.

$sel:sSHPublicKey:SendSerialConsoleSSHPublicKey', sendSerialConsoleSSHPublicKey_sSHPublicKey - The public key material. To use the public key, you must have the matching private key. For information about the supported key formats and lengths, see Requirements for key pairs in the Amazon EC2 User Guide.

data SendSerialConsoleSSHPublicKeyResponse

See: newSendSerialConsoleSSHPublicKeyResponse smart constructor.

Constructors

SendSerialConsoleSSHPublicKeyResponse' (Maybe Text) (Maybe Bool) Int

newSendSerialConsoleSSHPublicKeyResponse

Arguments

:: Int
$sel:httpStatus:SendSerialConsoleSSHPublicKeyResponse'

-> SendSerialConsoleSSHPublicKeyResponse

Create a value of SendSerialConsoleSSHPublicKeyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:SendSerialConsoleSSHPublicKeyResponse', sendSerialConsoleSSHPublicKeyResponse_requestId - The ID of the request. Please provide this ID when contacting AWS Support for assistance.

$sel:success:SendSerialConsoleSSHPublicKeyResponse', sendSerialConsoleSSHPublicKeyResponse_success - Is true if the request succeeds and an error otherwise.

$sel:httpStatus:SendSerialConsoleSSHPublicKeyResponse', sendSerialConsoleSSHPublicKeyResponse_httpStatus - The response's http status code.

Types