Network.AWS.EC2InstanceConnect.Types

Description

Service Configuration

defaultService :: Service

API version 2018-04-02 of the Amazon EC2 Instance Connect SDK configuration.

Errors

_SerialConsoleSessionUnavailableException :: AsError a => Getting (First ServiceError) a ServiceError

Unable to start a serial console session. Please try again.

_AuthException :: AsError a => Getting (First ServiceError) a ServiceError

Either your AWS credentials are not valid or you do not have access to the EC2 instance.

_SerialConsoleSessionLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The instance currently has 1 active serial console session. Only 1 session is supported at a time.

_InvalidArgsException :: AsError a => Getting (First ServiceError) a ServiceError

One of the parameters is not valid.

_SerialConsoleAccessDisabledException :: AsError a => Getting (First ServiceError) a ServiceError

Your account is not authorized to use the EC2 Serial Console. To authorize your account, run the EnableSerialConsoleAccess API. For more information, see EnableSerialConsoleAccess in the Amazon EC2 API Reference.

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The requests were made too frequently and have been throttled. Wait a while and try again. To increase the limit on your request frequency, contact AWS Support.

_ServiceException :: AsError a => Getting (First ServiceError) a ServiceError

The service encountered an error. Follow the instructions in the error message and try again.

_EC2InstanceTypeInvalidException :: AsError a => Getting (First ServiceError) a ServiceError

The instance type is not supported for connecting via the serial console. Only Nitro instance types are currently supported.

_EC2InstanceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified instance was not found.