Network.AWS.ECR.BatchGetImage

Description

Gets detailed information for an image. Images are specified with either an imageTag or imageDigest.

When an image is pulled, the BatchGetImage API is called once to retrieve the image manifest.

Creating a Request

data BatchGetImage

See: newBatchGetImage smart constructor.

Constructors

BatchGetImage'

Fields

registryId :: Maybe Text

The Amazon Web Services account ID associated with the registry that contains the images to describe. If you do not specify a registry, the default registry is assumed.

acceptedMediaTypes :: Maybe (NonEmpty Text)

The accepted media types for the request.

Valid values: application/vnd.docker.distribution.manifest.v1+json | application/vnd.docker.distribution.manifest.v2+json | application/vnd.oci.image.manifest.v1+json

repositoryName :: Text

The repository that contains the images to describe.

imageIds :: [ImageIdentifier]

A list of image ID references that correspond to images to describe. The format of the imageIds reference is imageTag=tag or imageDigest=digest.

newBatchGetImage

Arguments

:: Text
$sel:repositoryName:BatchGetImage'

-> BatchGetImage

Create a value of BatchGetImage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:registryId:BatchGetImage', batchGetImage_registryId - The Amazon Web Services account ID associated with the registry that contains the images to describe. If you do not specify a registry, the default registry is assumed.

$sel:acceptedMediaTypes:BatchGetImage', batchGetImage_acceptedMediaTypes - The accepted media types for the request.

Valid values: application/vnd.docker.distribution.manifest.v1+json | application/vnd.docker.distribution.manifest.v2+json | application/vnd.oci.image.manifest.v1+json

$sel:repositoryName:BatchGetImage', batchGetImage_repositoryName - The repository that contains the images to describe.

$sel:imageIds:BatchGetImage', batchGetImage_imageIds - A list of image ID references that correspond to images to describe. The format of the imageIds reference is imageTag=tag or imageDigest=digest.

Request Lenses

batchGetImage_registryId :: Lens' BatchGetImage (Maybe Text)

The Amazon Web Services account ID associated with the registry that contains the images to describe. If you do not specify a registry, the default registry is assumed.

batchGetImage_acceptedMediaTypes :: Lens' BatchGetImage (Maybe (NonEmpty Text))

The accepted media types for the request.

Valid values: application/vnd.docker.distribution.manifest.v1+json | application/vnd.docker.distribution.manifest.v2+json | application/vnd.oci.image.manifest.v1+json

batchGetImage_repositoryName :: Lens' BatchGetImage Text

The repository that contains the images to describe.

batchGetImage_imageIds :: Lens' BatchGetImage [ImageIdentifier]

A list of image ID references that correspond to images to describe. The format of the imageIds reference is imageTag=tag or imageDigest=digest.

Destructuring the Response

data BatchGetImageResponse

See: newBatchGetImageResponse smart constructor.

Constructors

BatchGetImageResponse'

Fields

images :: Maybe [Image]

A list of image objects corresponding to the image references in the request.

failures :: Maybe [ImageFailure]

Any failures associated with the call.

httpStatus :: Int

The response's http status code.

newBatchGetImageResponse

Arguments

:: Int
$sel:httpStatus:BatchGetImageResponse'

-> BatchGetImageResponse

Create a value of BatchGetImageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:images:BatchGetImageResponse', batchGetImageResponse_images - A list of image objects corresponding to the image references in the request.

$sel:failures:BatchGetImageResponse', batchGetImageResponse_failures - Any failures associated with the call.

$sel:httpStatus:BatchGetImageResponse', batchGetImageResponse_httpStatus - The response's http status code.

Response Lenses

batchGetImageResponse_images :: Lens' BatchGetImageResponse (Maybe [Image])

A list of image objects corresponding to the image references in the request.

batchGetImageResponse_failures :: Lens' BatchGetImageResponse (Maybe [ImageFailure])

Any failures associated with the call.

batchGetImageResponse_httpStatus :: Lens' BatchGetImageResponse Int

The response's http status code.