Network.AWS.ECR.DeleteRepositoryPolicy

Description

Deletes the repository policy associated with the specified repository.

Creating a Request

data DeleteRepositoryPolicy

See: newDeleteRepositoryPolicy smart constructor.

Constructors

DeleteRepositoryPolicy'

Fields

registryId :: Maybe Text

The Amazon Web Services account ID associated with the registry that contains the repository policy to delete. If you do not specify a registry, the default registry is assumed.

repositoryName :: Text

The name of the repository that is associated with the repository policy to delete.

newDeleteRepositoryPolicy

Arguments

:: Text
$sel:repositoryName:DeleteRepositoryPolicy'

-> DeleteRepositoryPolicy

Create a value of DeleteRepositoryPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:registryId:DeleteRepositoryPolicy', deleteRepositoryPolicy_registryId - The Amazon Web Services account ID associated with the registry that contains the repository policy to delete. If you do not specify a registry, the default registry is assumed.

$sel:repositoryName:DeleteRepositoryPolicy', deleteRepositoryPolicy_repositoryName - The name of the repository that is associated with the repository policy to delete.

Request Lenses

deleteRepositoryPolicy_registryId :: Lens' DeleteRepositoryPolicy (Maybe Text)

The Amazon Web Services account ID associated with the registry that contains the repository policy to delete. If you do not specify a registry, the default registry is assumed.

deleteRepositoryPolicy_repositoryName :: Lens' DeleteRepositoryPolicy Text

The name of the repository that is associated with the repository policy to delete.

Destructuring the Response

data DeleteRepositoryPolicyResponse

See: newDeleteRepositoryPolicyResponse smart constructor.

Constructors

DeleteRepositoryPolicyResponse'

Fields

registryId :: Maybe Text

The registry ID associated with the request.

repositoryName :: Maybe Text

The repository name associated with the request.

policyText :: Maybe Text

The JSON repository policy that was deleted from the repository.

httpStatus :: Int

The response's http status code.

newDeleteRepositoryPolicyResponse

Arguments

:: Int
$sel:httpStatus:DeleteRepositoryPolicyResponse'

-> DeleteRepositoryPolicyResponse

Create a value of DeleteRepositoryPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:registryId:DeleteRepositoryPolicy', deleteRepositoryPolicyResponse_registryId - The registry ID associated with the request.

$sel:repositoryName:DeleteRepositoryPolicy', deleteRepositoryPolicyResponse_repositoryName - The repository name associated with the request.

$sel:policyText:DeleteRepositoryPolicyResponse', deleteRepositoryPolicyResponse_policyText - The JSON repository policy that was deleted from the repository.

$sel:httpStatus:DeleteRepositoryPolicyResponse', deleteRepositoryPolicyResponse_httpStatus - The response's http status code.

Response Lenses

deleteRepositoryPolicyResponse_registryId :: Lens' DeleteRepositoryPolicyResponse (Maybe Text)

The registry ID associated with the request.

deleteRepositoryPolicyResponse_repositoryName :: Lens' DeleteRepositoryPolicyResponse (Maybe Text)

The repository name associated with the request.

deleteRepositoryPolicyResponse_policyText :: Lens' DeleteRepositoryPolicyResponse (Maybe Text)

The JSON repository policy that was deleted from the repository.

deleteRepositoryPolicyResponse_httpStatus :: Lens' DeleteRepositoryPolicyResponse Int

The response's http status code.