Network.AWS.ECR.DescribeImageScanFindings

Description

Returns the scan findings for the specified image.

This operation returns paginated results.

Creating a Request

data DescribeImageScanFindings

See: newDescribeImageScanFindings smart constructor.

Constructors

DescribeImageScanFindings'

Fields

registryId :: Maybe Text

The Amazon Web Services account ID associated with the registry that contains the repository in which to describe the image scan findings for. If you do not specify a registry, the default registry is assumed.

nextToken :: Maybe Text

The nextToken value returned from a previous paginated DescribeImageScanFindings request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return.

maxResults :: Maybe Natural

The maximum number of image scan results returned by DescribeImageScanFindings in paginated output. When this parameter is used, DescribeImageScanFindings only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another DescribeImageScanFindings request with the returned nextToken value. This value can be between 1 and 1000. If this parameter is not used, then DescribeImageScanFindings returns up to 100 results and a nextToken value, if applicable.

repositoryName :: Text

The repository for the image for which to describe the scan findings.

imageId :: ImageIdentifier

newDescribeImageScanFindings

Arguments

:: Text
$sel:repositoryName:DescribeImageScanFindings'

-> ImageIdentifier
$sel:imageId:DescribeImageScanFindings'

-> DescribeImageScanFindings

Create a value of DescribeImageScanFindings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:registryId:DescribeImageScanFindings', describeImageScanFindings_registryId - The Amazon Web Services account ID associated with the registry that contains the repository in which to describe the image scan findings for. If you do not specify a registry, the default registry is assumed.

$sel:nextToken:DescribeImageScanFindings', describeImageScanFindings_nextToken - The nextToken value returned from a previous paginated DescribeImageScanFindings request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return.

$sel:maxResults:DescribeImageScanFindings', describeImageScanFindings_maxResults - The maximum number of image scan results returned by DescribeImageScanFindings in paginated output. When this parameter is used, DescribeImageScanFindings only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another DescribeImageScanFindings request with the returned nextToken value. This value can be between 1 and 1000. If this parameter is not used, then DescribeImageScanFindings returns up to 100 results and a nextToken value, if applicable.

$sel:repositoryName:DescribeImageScanFindings', describeImageScanFindings_repositoryName - The repository for the image for which to describe the scan findings.

$sel:imageId:DescribeImageScanFindings', describeImageScanFindings_imageId - Undocumented member.

Request Lenses

describeImageScanFindings_registryId :: Lens' DescribeImageScanFindings (Maybe Text)

The Amazon Web Services account ID associated with the registry that contains the repository in which to describe the image scan findings for. If you do not specify a registry, the default registry is assumed.

describeImageScanFindings_nextToken :: Lens' DescribeImageScanFindings (Maybe Text)

The nextToken value returned from a previous paginated DescribeImageScanFindings request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return.

describeImageScanFindings_maxResults :: Lens' DescribeImageScanFindings (Maybe Natural)

The maximum number of image scan results returned by DescribeImageScanFindings in paginated output. When this parameter is used, DescribeImageScanFindings only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another DescribeImageScanFindings request with the returned nextToken value. This value can be between 1 and 1000. If this parameter is not used, then DescribeImageScanFindings returns up to 100 results and a nextToken value, if applicable.

describeImageScanFindings_repositoryName :: Lens' DescribeImageScanFindings Text

The repository for the image for which to describe the scan findings.

describeImageScanFindings_imageId :: Lens' DescribeImageScanFindings ImageIdentifier

Undocumented member.

Destructuring the Response

data DescribeImageScanFindingsResponse

See: newDescribeImageScanFindingsResponse smart constructor.

Constructors

DescribeImageScanFindingsResponse'

Fields

registryId :: Maybe Text

The registry ID associated with the request.

imageScanFindings :: Maybe ImageScanFindings

The information contained in the image scan findings.

imageScanStatus :: Maybe ImageScanStatus

The current state of the scan.

nextToken :: Maybe Text

The nextToken value to include in a future DescribeImageScanFindings request. When the results of a DescribeImageScanFindings request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

imageId :: Maybe ImageIdentifier

repositoryName :: Maybe Text

The repository name associated with the request.

httpStatus :: Int

The response's http status code.

newDescribeImageScanFindingsResponse

Arguments

:: Int
$sel:httpStatus:DescribeImageScanFindingsResponse'

-> DescribeImageScanFindingsResponse

Create a value of DescribeImageScanFindingsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:registryId:DescribeImageScanFindings', describeImageScanFindingsResponse_registryId - The registry ID associated with the request.

$sel:imageScanFindings:DescribeImageScanFindingsResponse', describeImageScanFindingsResponse_imageScanFindings - The information contained in the image scan findings.

$sel:imageScanStatus:DescribeImageScanFindingsResponse', describeImageScanFindingsResponse_imageScanStatus - The current state of the scan.

$sel:nextToken:DescribeImageScanFindings', describeImageScanFindingsResponse_nextToken - The nextToken value to include in a future DescribeImageScanFindings request. When the results of a DescribeImageScanFindings request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

$sel:imageId:DescribeImageScanFindings', describeImageScanFindingsResponse_imageId - Undocumented member.

$sel:repositoryName:DescribeImageScanFindings', describeImageScanFindingsResponse_repositoryName - The repository name associated with the request.

$sel:httpStatus:DescribeImageScanFindingsResponse', describeImageScanFindingsResponse_httpStatus - The response's http status code.

Response Lenses

describeImageScanFindingsResponse_registryId :: Lens' DescribeImageScanFindingsResponse (Maybe Text)

The registry ID associated with the request.

describeImageScanFindingsResponse_imageScanFindings :: Lens' DescribeImageScanFindingsResponse (Maybe ImageScanFindings)

The information contained in the image scan findings.

describeImageScanFindingsResponse_imageScanStatus :: Lens' DescribeImageScanFindingsResponse (Maybe ImageScanStatus)

The current state of the scan.

describeImageScanFindingsResponse_nextToken :: Lens' DescribeImageScanFindingsResponse (Maybe Text)

The nextToken value to include in a future DescribeImageScanFindings request. When the results of a DescribeImageScanFindings request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

describeImageScanFindingsResponse_imageId :: Lens' DescribeImageScanFindingsResponse (Maybe ImageIdentifier)

Undocumented member.

describeImageScanFindingsResponse_repositoryName :: Lens' DescribeImageScanFindingsResponse (Maybe Text)

The repository name associated with the request.

describeImageScanFindingsResponse_httpStatus :: Lens' DescribeImageScanFindingsResponse Int

The response's http status code.