Network.AWS.ECR.DescribeImages

Description

Returns metadata about the images in a repository.

Beginning with Docker version 1.9, the Docker client compresses image layers before pushing them to a V2 Docker registry. The output of the docker images command shows the uncompressed image size, so it may return a larger image size than the image sizes returned by DescribeImages.

This operation returns paginated results.

Creating a Request

data DescribeImages

See: newDescribeImages smart constructor.

Constructors

DescribeImages'

Fields

nextToken :: Maybe Text

The nextToken value returned from a previous paginated DescribeImages request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return. This option cannot be used when you specify images with imageIds.

imageIds :: Maybe [ImageIdentifier]

The list of image IDs for the requested repository.

maxResults :: Maybe Natural

The maximum number of repository results returned by DescribeImages in paginated output. When this parameter is used, DescribeImages only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another DescribeImages request with the returned nextToken value. This value can be between 1 and 1000. If this parameter is not used, then DescribeImages returns up to 100 results and a nextToken value, if applicable. This option cannot be used when you specify images with imageIds.

registryId :: Maybe Text

The Amazon Web Services account ID associated with the registry that contains the repository in which to describe images. If you do not specify a registry, the default registry is assumed.

filter' :: Maybe DescribeImagesFilter

The filter key and value with which to filter your DescribeImages results.

repositoryName :: Text

The repository that contains the images to describe.

newDescribeImages

Arguments

:: Text
$sel:repositoryName:DescribeImages'

-> DescribeImages

Create a value of DescribeImages with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeImages', describeImages_nextToken - The nextToken value returned from a previous paginated DescribeImages request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return. This option cannot be used when you specify images with imageIds.

$sel:imageIds:DescribeImages', describeImages_imageIds - The list of image IDs for the requested repository.

$sel:maxResults:DescribeImages', describeImages_maxResults - The maximum number of repository results returned by DescribeImages in paginated output. When this parameter is used, DescribeImages only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another DescribeImages request with the returned nextToken value. This value can be between 1 and 1000. If this parameter is not used, then DescribeImages returns up to 100 results and a nextToken value, if applicable. This option cannot be used when you specify images with imageIds.

$sel:registryId:DescribeImages', describeImages_registryId - The Amazon Web Services account ID associated with the registry that contains the repository in which to describe images. If you do not specify a registry, the default registry is assumed.

$sel:filter':DescribeImages', describeImages_filter - The filter key and value with which to filter your DescribeImages results.

$sel:repositoryName:DescribeImages', describeImages_repositoryName - The repository that contains the images to describe.

Request Lenses

describeImages_nextToken :: Lens' DescribeImages (Maybe Text)

The nextToken value returned from a previous paginated DescribeImages request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return. This option cannot be used when you specify images with imageIds.

describeImages_imageIds :: Lens' DescribeImages (Maybe [ImageIdentifier])

The list of image IDs for the requested repository.

describeImages_maxResults :: Lens' DescribeImages (Maybe Natural)

The maximum number of repository results returned by DescribeImages in paginated output. When this parameter is used, DescribeImages only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another DescribeImages request with the returned nextToken value. This value can be between 1 and 1000. If this parameter is not used, then DescribeImages returns up to 100 results and a nextToken value, if applicable. This option cannot be used when you specify images with imageIds.

describeImages_registryId :: Lens' DescribeImages (Maybe Text)

The Amazon Web Services account ID associated with the registry that contains the repository in which to describe images. If you do not specify a registry, the default registry is assumed.

describeImages_filter :: Lens' DescribeImages (Maybe DescribeImagesFilter)

The filter key and value with which to filter your DescribeImages results.

describeImages_repositoryName :: Lens' DescribeImages Text

The repository that contains the images to describe.

Destructuring the Response

data DescribeImagesResponse

See: newDescribeImagesResponse smart constructor.

Constructors

DescribeImagesResponse'

Fields

nextToken :: Maybe Text

The nextToken value to include in a future DescribeImages request. When the results of a DescribeImages request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

imageDetails :: Maybe [ImageDetail]

A list of ImageDetail objects that contain data about the image.

httpStatus :: Int

The response's http status code.

newDescribeImagesResponse

Arguments

:: Int
$sel:httpStatus:DescribeImagesResponse'

-> DescribeImagesResponse

Create a value of DescribeImagesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeImages', describeImagesResponse_nextToken - The nextToken value to include in a future DescribeImages request. When the results of a DescribeImages request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

$sel:imageDetails:DescribeImagesResponse', describeImagesResponse_imageDetails - A list of ImageDetail objects that contain data about the image.

$sel:httpStatus:DescribeImagesResponse', describeImagesResponse_httpStatus - The response's http status code.

Response Lenses

describeImagesResponse_nextToken :: Lens' DescribeImagesResponse (Maybe Text)

The nextToken value to include in a future DescribeImages request. When the results of a DescribeImages request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

describeImagesResponse_imageDetails :: Lens' DescribeImagesResponse (Maybe [ImageDetail])

A list of ImageDetail objects that contain data about the image.

describeImagesResponse_httpStatus :: Lens' DescribeImagesResponse Int

The response's http status code.