Network.AWS.ECR.GetDownloadUrlForLayer

Description

Retrieves the pre-signed Amazon S3 download URL corresponding to an image layer. You can only get URLs for image layers that are referenced in an image.

When an image is pulled, the GetDownloadUrlForLayer API is called once per image layer that is not already cached.

This operation is used by the Amazon ECR proxy and is not generally used by customers for pulling and pushing images. In most cases, you should use the docker CLI to pull, tag, and push images.

Creating a Request

data GetDownloadUrlForLayer

See: newGetDownloadUrlForLayer smart constructor.

Constructors

GetDownloadUrlForLayer'

Fields

registryId :: Maybe Text

The Amazon Web Services account ID associated with the registry that contains the image layer to download. If you do not specify a registry, the default registry is assumed.

repositoryName :: Text

The name of the repository that is associated with the image layer to download.

layerDigest :: Text

The digest of the image layer to download.

newGetDownloadUrlForLayer

Arguments

:: Text
$sel:repositoryName:GetDownloadUrlForLayer'

-> Text
$sel:layerDigest:GetDownloadUrlForLayer'

-> GetDownloadUrlForLayer

Create a value of GetDownloadUrlForLayer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:registryId:GetDownloadUrlForLayer', getDownloadUrlForLayer_registryId - The Amazon Web Services account ID associated with the registry that contains the image layer to download. If you do not specify a registry, the default registry is assumed.

$sel:repositoryName:GetDownloadUrlForLayer', getDownloadUrlForLayer_repositoryName - The name of the repository that is associated with the image layer to download.

$sel:layerDigest:GetDownloadUrlForLayer', getDownloadUrlForLayer_layerDigest - The digest of the image layer to download.

Request Lenses

getDownloadUrlForLayer_registryId :: Lens' GetDownloadUrlForLayer (Maybe Text)

The Amazon Web Services account ID associated with the registry that contains the image layer to download. If you do not specify a registry, the default registry is assumed.

getDownloadUrlForLayer_repositoryName :: Lens' GetDownloadUrlForLayer Text

The name of the repository that is associated with the image layer to download.

getDownloadUrlForLayer_layerDigest :: Lens' GetDownloadUrlForLayer Text

The digest of the image layer to download.

Destructuring the Response

data GetDownloadUrlForLayerResponse

See: newGetDownloadUrlForLayerResponse smart constructor.

Constructors

GetDownloadUrlForLayerResponse'

Fields

layerDigest :: Maybe Text

The digest of the image layer to download.

downloadUrl :: Maybe Text

The pre-signed Amazon S3 download URL for the requested layer.

httpStatus :: Int

The response's http status code.

newGetDownloadUrlForLayerResponse

Arguments

:: Int
$sel:httpStatus:GetDownloadUrlForLayerResponse'

-> GetDownloadUrlForLayerResponse

Create a value of GetDownloadUrlForLayerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:layerDigest:GetDownloadUrlForLayer', getDownloadUrlForLayerResponse_layerDigest - The digest of the image layer to download.

$sel:downloadUrl:GetDownloadUrlForLayerResponse', getDownloadUrlForLayerResponse_downloadUrl - The pre-signed Amazon S3 download URL for the requested layer.

$sel:httpStatus:GetDownloadUrlForLayerResponse', getDownloadUrlForLayerResponse_httpStatus - The response's http status code.

Response Lenses

getDownloadUrlForLayerResponse_layerDigest :: Lens' GetDownloadUrlForLayerResponse (Maybe Text)

The digest of the image layer to download.

getDownloadUrlForLayerResponse_downloadUrl :: Lens' GetDownloadUrlForLayerResponse (Maybe Text)

The pre-signed Amazon S3 download URL for the requested layer.

getDownloadUrlForLayerResponse_httpStatus :: Lens' GetDownloadUrlForLayerResponse Int

The response's http status code.