Network.AWS.ECR.GetLifecyclePolicy

Description

Retrieves the lifecycle policy for the specified repository.

Creating a Request

data GetLifecyclePolicy

See: newGetLifecyclePolicy smart constructor.

Constructors

GetLifecyclePolicy'

Fields

registryId :: Maybe Text

The Amazon Web Services account ID associated with the registry that contains the repository. If you do not specify a registry, the default registry is assumed.

repositoryName :: Text

The name of the repository.

newGetLifecyclePolicy

Arguments

:: Text
$sel:repositoryName:GetLifecyclePolicy'

-> GetLifecyclePolicy

Create a value of GetLifecyclePolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:registryId:GetLifecyclePolicy', getLifecyclePolicy_registryId - The Amazon Web Services account ID associated with the registry that contains the repository. If you do not specify a registry, the default registry is assumed.

$sel:repositoryName:GetLifecyclePolicy', getLifecyclePolicy_repositoryName - The name of the repository.

Request Lenses

getLifecyclePolicy_registryId :: Lens' GetLifecyclePolicy (Maybe Text)

The Amazon Web Services account ID associated with the registry that contains the repository. If you do not specify a registry, the default registry is assumed.

getLifecyclePolicy_repositoryName :: Lens' GetLifecyclePolicy Text

The name of the repository.

Destructuring the Response

data GetLifecyclePolicyResponse

See: newGetLifecyclePolicyResponse smart constructor.

Constructors

GetLifecyclePolicyResponse'

Fields

registryId :: Maybe Text

The registry ID associated with the request.

lastEvaluatedAt :: Maybe POSIX

The time stamp of the last time that the lifecycle policy was run.

lifecyclePolicyText :: Maybe Text

The JSON lifecycle policy text.

repositoryName :: Maybe Text

The repository name associated with the request.

httpStatus :: Int

The response's http status code.

newGetLifecyclePolicyResponse

Arguments

:: Int
$sel:httpStatus:GetLifecyclePolicyResponse'

-> GetLifecyclePolicyResponse

Create a value of GetLifecyclePolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:registryId:GetLifecyclePolicy', getLifecyclePolicyResponse_registryId - The registry ID associated with the request.

$sel:lastEvaluatedAt:GetLifecyclePolicyResponse', getLifecyclePolicyResponse_lastEvaluatedAt - The time stamp of the last time that the lifecycle policy was run.

$sel:lifecyclePolicyText:GetLifecyclePolicyResponse', getLifecyclePolicyResponse_lifecyclePolicyText - The JSON lifecycle policy text.

$sel:repositoryName:GetLifecyclePolicy', getLifecyclePolicyResponse_repositoryName - The repository name associated with the request.

$sel:httpStatus:GetLifecyclePolicyResponse', getLifecyclePolicyResponse_httpStatus - The response's http status code.

Response Lenses

getLifecyclePolicyResponse_registryId :: Lens' GetLifecyclePolicyResponse (Maybe Text)

The registry ID associated with the request.

getLifecyclePolicyResponse_lastEvaluatedAt :: Lens' GetLifecyclePolicyResponse (Maybe UTCTime)

The time stamp of the last time that the lifecycle policy was run.

getLifecyclePolicyResponse_lifecyclePolicyText :: Lens' GetLifecyclePolicyResponse (Maybe Text)

The JSON lifecycle policy text.

getLifecyclePolicyResponse_repositoryName :: Lens' GetLifecyclePolicyResponse (Maybe Text)

The repository name associated with the request.

getLifecyclePolicyResponse_httpStatus :: Lens' GetLifecyclePolicyResponse Int

The response's http status code.