Network.AWS.ECR.Lens

Description

Operations

UploadLayerPart

uploadLayerPart_registryId :: Lens' UploadLayerPart (Maybe Text)

The Amazon Web Services account ID associated with the registry to which you are uploading layer parts. If you do not specify a registry, the default registry is assumed.

uploadLayerPart_repositoryName :: Lens' UploadLayerPart Text

The name of the repository to which you are uploading layer parts.

uploadLayerPart_uploadId :: Lens' UploadLayerPart Text

The upload ID from a previous InitiateLayerUpload operation to associate with the layer part upload.

uploadLayerPart_partFirstByte :: Lens' UploadLayerPart Natural

The position of the first byte of the layer part witin the overall image layer.

uploadLayerPart_partLastByte :: Lens' UploadLayerPart Natural

The position of the last byte of the layer part within the overall image layer.

uploadLayerPart_layerPartBlob :: Lens' UploadLayerPart ByteString

The base64-encoded layer part payload.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

uploadLayerPartResponse_uploadId :: Lens' UploadLayerPartResponse (Maybe Text)

The upload ID associated with the request.

uploadLayerPartResponse_registryId :: Lens' UploadLayerPartResponse (Maybe Text)

The registry ID associated with the request.

uploadLayerPartResponse_repositoryName :: Lens' UploadLayerPartResponse (Maybe Text)

The repository name associated with the request.

uploadLayerPartResponse_lastByteReceived :: Lens' UploadLayerPartResponse (Maybe Natural)

The integer value of the last byte received in the request.

uploadLayerPartResponse_httpStatus :: Lens' UploadLayerPartResponse Int

The response's http status code.

PutLifecyclePolicy

putLifecyclePolicy_registryId :: Lens' PutLifecyclePolicy (Maybe Text)

The Amazon Web Services account ID associated with the registry that contains the repository. If you do not specify a registry, the default registry is assumed.

putLifecyclePolicy_repositoryName :: Lens' PutLifecyclePolicy Text

The name of the repository to receive the policy.

putLifecyclePolicy_lifecyclePolicyText :: Lens' PutLifecyclePolicy Text

The JSON repository policy text to apply to the repository.

putLifecyclePolicyResponse_registryId :: Lens' PutLifecyclePolicyResponse (Maybe Text)

The registry ID associated with the request.

putLifecyclePolicyResponse_repositoryName :: Lens' PutLifecyclePolicyResponse (Maybe Text)

The repository name associated with the request.

putLifecyclePolicyResponse_lifecyclePolicyText :: Lens' PutLifecyclePolicyResponse (Maybe Text)

The JSON repository policy text.

putLifecyclePolicyResponse_httpStatus :: Lens' PutLifecyclePolicyResponse Int

The response's http status code.

PutRegistryPolicy

putRegistryPolicy_policyText :: Lens' PutRegistryPolicy Text

The JSON policy text to apply to your registry. The policy text follows the same format as IAM policy text. For more information, see Registry permissions in the Amazon Elastic Container Registry User Guide.

putRegistryPolicyResponse_registryId :: Lens' PutRegistryPolicyResponse (Maybe Text)

The registry ID.

putRegistryPolicyResponse_policyText :: Lens' PutRegistryPolicyResponse (Maybe Text)

The JSON policy text for your registry.

putRegistryPolicyResponse_httpStatus :: Lens' PutRegistryPolicyResponse Int

The response's http status code.

StartLifecyclePolicyPreview

startLifecyclePolicyPreview_registryId :: Lens' StartLifecyclePolicyPreview (Maybe Text)

The Amazon Web Services account ID associated with the registry that contains the repository. If you do not specify a registry, the default registry is assumed.

startLifecyclePolicyPreview_lifecyclePolicyText :: Lens' StartLifecyclePolicyPreview (Maybe Text)

The policy to be evaluated against. If you do not specify a policy, the current policy for the repository is used.

startLifecyclePolicyPreview_repositoryName :: Lens' StartLifecyclePolicyPreview Text

The name of the repository to be evaluated.

startLifecyclePolicyPreviewResponse_status :: Lens' StartLifecyclePolicyPreviewResponse (Maybe LifecyclePolicyPreviewStatus)

The status of the lifecycle policy preview request.

startLifecyclePolicyPreviewResponse_registryId :: Lens' StartLifecyclePolicyPreviewResponse (Maybe Text)

The registry ID associated with the request.

startLifecyclePolicyPreviewResponse_repositoryName :: Lens' StartLifecyclePolicyPreviewResponse (Maybe Text)

The repository name associated with the request.

startLifecyclePolicyPreviewResponse_lifecyclePolicyText :: Lens' StartLifecyclePolicyPreviewResponse (Maybe Text)

The JSON repository policy text.

startLifecyclePolicyPreviewResponse_httpStatus :: Lens' StartLifecyclePolicyPreviewResponse Int

The response's http status code.

DescribeRepositories

describeRepositories_nextToken :: Lens' DescribeRepositories (Maybe Text)

The nextToken value returned from a previous paginated DescribeRepositories request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return. This option cannot be used when you specify repositories with repositoryNames.

This token should be treated as an opaque identifier that is only used to retrieve the next items in a list and not for other programmatic purposes.

describeRepositories_repositoryNames :: Lens' DescribeRepositories (Maybe (NonEmpty Text))

A list of repositories to describe. If this parameter is omitted, then all repositories in a registry are described.

describeRepositories_maxResults :: Lens' DescribeRepositories (Maybe Natural)

The maximum number of repository results returned by DescribeRepositories in paginated output. When this parameter is used, DescribeRepositories only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another DescribeRepositories request with the returned nextToken value. This value can be between 1 and 1000. If this parameter is not used, then DescribeRepositories returns up to 100 results and a nextToken value, if applicable. This option cannot be used when you specify repositories with repositoryNames.

describeRepositories_registryId :: Lens' DescribeRepositories (Maybe Text)

The Amazon Web Services account ID associated with the registry that contains the repositories to be described. If you do not specify a registry, the default registry is assumed.

describeRepositoriesResponse_nextToken :: Lens' DescribeRepositoriesResponse (Maybe Text)

The nextToken value to include in a future DescribeRepositories request. When the results of a DescribeRepositories request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

describeRepositoriesResponse_repositories :: Lens' DescribeRepositoriesResponse (Maybe [Repository])

A list of repository objects corresponding to valid repositories.

describeRepositoriesResponse_httpStatus :: Lens' DescribeRepositoriesResponse Int

The response's http status code.

PutImage

putImage_imageDigest :: Lens' PutImage (Maybe Text)

The image digest of the image manifest corresponding to the image.

putImage_registryId :: Lens' PutImage (Maybe Text)

The Amazon Web Services account ID associated with the registry that contains the repository in which to put the image. If you do not specify a registry, the default registry is assumed.

putImage_imageTag :: Lens' PutImage (Maybe Text)

The tag to associate with the image. This parameter is required for images that use the Docker Image Manifest V2 Schema 2 or Open Container Initiative (OCI) formats.

putImage_imageManifestMediaType :: Lens' PutImage (Maybe Text)

The media type of the image manifest. If you push an image manifest that does not contain the mediaType field, you must specify the imageManifestMediaType in the request.

putImage_repositoryName :: Lens' PutImage Text

The name of the repository in which to put the image.

putImage_imageManifest :: Lens' PutImage Text

The image manifest corresponding to the image to be uploaded.

putImageResponse_image :: Lens' PutImageResponse (Maybe Image)

Details of the image uploaded.

putImageResponse_httpStatus :: Lens' PutImageResponse Int

The response's http status code.

ListImages

listImages_nextToken :: Lens' ListImages (Maybe Text)

The nextToken value returned from a previous paginated ListImages request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return.

This token should be treated as an opaque identifier that is only used to retrieve the next items in a list and not for other programmatic purposes.

listImages_maxResults :: Lens' ListImages (Maybe Natural)

The maximum number of image results returned by ListImages in paginated output. When this parameter is used, ListImages only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListImages request with the returned nextToken value. This value can be between 1 and 1000. If this parameter is not used, then ListImages returns up to 100 results and a nextToken value, if applicable.

listImages_registryId :: Lens' ListImages (Maybe Text)

The Amazon Web Services account ID associated with the registry that contains the repository in which to list images. If you do not specify a registry, the default registry is assumed.

listImages_filter :: Lens' ListImages (Maybe ListImagesFilter)

The filter key and value with which to filter your ListImages results.

listImages_repositoryName :: Lens' ListImages Text

The repository with image IDs to be listed.

listImagesResponse_nextToken :: Lens' ListImagesResponse (Maybe Text)

The nextToken value to include in a future ListImages request. When the results of a ListImages request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

listImagesResponse_imageIds :: Lens' ListImagesResponse (Maybe [ImageIdentifier])

The list of image IDs for the requested repository.

listImagesResponse_httpStatus :: Lens' ListImagesResponse Int

The response's http status code.

GetRegistryPolicy

getRegistryPolicyResponse_registryId :: Lens' GetRegistryPolicyResponse (Maybe Text)

The ID of the registry.

getRegistryPolicyResponse_policyText :: Lens' GetRegistryPolicyResponse (Maybe Text)

The JSON text of the permissions policy for a registry.

getRegistryPolicyResponse_httpStatus :: Lens' GetRegistryPolicyResponse Int

The response's http status code.

InitiateLayerUpload

initiateLayerUpload_registryId :: Lens' InitiateLayerUpload (Maybe Text)

The Amazon Web Services account ID associated with the registry to which you intend to upload layers. If you do not specify a registry, the default registry is assumed.

initiateLayerUpload_repositoryName :: Lens' InitiateLayerUpload Text

The name of the repository to which you intend to upload layers.

initiateLayerUploadResponse_uploadId :: Lens' InitiateLayerUploadResponse (Maybe Text)

The upload ID for the layer upload. This parameter is passed to further UploadLayerPart and CompleteLayerUpload operations.

initiateLayerUploadResponse_partSize :: Lens' InitiateLayerUploadResponse (Maybe Natural)

The size, in bytes, that Amazon ECR expects future layer part uploads to be.

initiateLayerUploadResponse_httpStatus :: Lens' InitiateLayerUploadResponse Int

The response's http status code.

DescribeImageScanFindings

describeImageScanFindings_nextToken :: Lens' DescribeImageScanFindings (Maybe Text)

The nextToken value returned from a previous paginated DescribeImageScanFindings request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return.

describeImageScanFindings_maxResults :: Lens' DescribeImageScanFindings (Maybe Natural)

The maximum number of image scan results returned by DescribeImageScanFindings in paginated output. When this parameter is used, DescribeImageScanFindings only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another DescribeImageScanFindings request with the returned nextToken value. This value can be between 1 and 1000. If this parameter is not used, then DescribeImageScanFindings returns up to 100 results and a nextToken value, if applicable.

describeImageScanFindings_registryId :: Lens' DescribeImageScanFindings (Maybe Text)

The Amazon Web Services account ID associated with the registry that contains the repository in which to describe the image scan findings for. If you do not specify a registry, the default registry is assumed.

describeImageScanFindings_repositoryName :: Lens' DescribeImageScanFindings Text

The repository for the image for which to describe the scan findings.

describeImageScanFindings_imageId :: Lens' DescribeImageScanFindings ImageIdentifier

Undocumented member.

describeImageScanFindingsResponse_imageScanStatus :: Lens' DescribeImageScanFindingsResponse (Maybe ImageScanStatus)

The current state of the scan.

describeImageScanFindingsResponse_nextToken :: Lens' DescribeImageScanFindingsResponse (Maybe Text)

The nextToken value to include in a future DescribeImageScanFindings request. When the results of a DescribeImageScanFindings request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

describeImageScanFindingsResponse_imageScanFindings :: Lens' DescribeImageScanFindingsResponse (Maybe ImageScanFindings)

The information contained in the image scan findings.

describeImageScanFindingsResponse_registryId :: Lens' DescribeImageScanFindingsResponse (Maybe Text)

The registry ID associated with the request.

describeImageScanFindingsResponse_repositoryName :: Lens' DescribeImageScanFindingsResponse (Maybe Text)

The repository name associated with the request.

describeImageScanFindingsResponse_imageId :: Lens' DescribeImageScanFindingsResponse (Maybe ImageIdentifier)

Undocumented member.

describeImageScanFindingsResponse_httpStatus :: Lens' DescribeImageScanFindingsResponse Int

The response's http status code.

DeleteRepositoryPolicy

deleteRepositoryPolicy_registryId :: Lens' DeleteRepositoryPolicy (Maybe Text)

The Amazon Web Services account ID associated with the registry that contains the repository policy to delete. If you do not specify a registry, the default registry is assumed.

deleteRepositoryPolicy_repositoryName :: Lens' DeleteRepositoryPolicy Text

The name of the repository that is associated with the repository policy to delete.

deleteRepositoryPolicyResponse_registryId :: Lens' DeleteRepositoryPolicyResponse (Maybe Text)

The registry ID associated with the request.

deleteRepositoryPolicyResponse_repositoryName :: Lens' DeleteRepositoryPolicyResponse (Maybe Text)

The repository name associated with the request.

deleteRepositoryPolicyResponse_policyText :: Lens' DeleteRepositoryPolicyResponse (Maybe Text)

The JSON repository policy that was deleted from the repository.

deleteRepositoryPolicyResponse_httpStatus :: Lens' DeleteRepositoryPolicyResponse Int

The response's http status code.

UntagResource

untagResource_resourceArn :: Lens' UntagResource Text

The Amazon Resource Name (ARN) of the resource from which to remove tags. Currently, the only supported resource is an Amazon ECR repository.

untagResource_tagKeys :: Lens' UntagResource [Text]

The keys of the tags to be removed.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The Amazon Resource Name (ARN) of the the resource to which to add tags. Currently, the only supported resource is an Amazon ECR repository.

tagResource_tags :: Lens' TagResource [Tag]

The tags to add to the resource. A tag is an array of key-value pairs. Tag keys can have a maximum character length of 128 characters, and tag values can have a maximum length of 256 characters.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

SetRepositoryPolicy

setRepositoryPolicy_registryId :: Lens' SetRepositoryPolicy (Maybe Text)

The Amazon Web Services account ID associated with the registry that contains the repository. If you do not specify a registry, the default registry is assumed.

setRepositoryPolicy_force :: Lens' SetRepositoryPolicy (Maybe Bool)

If the policy you are attempting to set on a repository policy would prevent you from setting another policy in the future, you must force the SetRepositoryPolicy operation. This is intended to prevent accidental repository lock outs.

setRepositoryPolicy_repositoryName :: Lens' SetRepositoryPolicy Text

The name of the repository to receive the policy.

setRepositoryPolicy_policyText :: Lens' SetRepositoryPolicy Text

The JSON repository policy text to apply to the repository. For more information, see Amazon ECR repository policies in the Amazon Elastic Container Registry User Guide.

setRepositoryPolicyResponse_registryId :: Lens' SetRepositoryPolicyResponse (Maybe Text)

The registry ID associated with the request.

setRepositoryPolicyResponse_repositoryName :: Lens' SetRepositoryPolicyResponse (Maybe Text)

The repository name associated with the request.

setRepositoryPolicyResponse_policyText :: Lens' SetRepositoryPolicyResponse (Maybe Text)

The JSON repository policy text applied to the repository.

setRepositoryPolicyResponse_httpStatus :: Lens' SetRepositoryPolicyResponse Int

The response's http status code.

DescribeRegistry

describeRegistryResponse_replicationConfiguration :: Lens' DescribeRegistryResponse (Maybe ReplicationConfiguration)

The replication configuration for the registry.

describeRegistryResponse_registryId :: Lens' DescribeRegistryResponse (Maybe Text)

The ID of the registry.

describeRegistryResponse_httpStatus :: Lens' DescribeRegistryResponse Int

The response's http status code.

BatchDeleteImage

batchDeleteImage_registryId :: Lens' BatchDeleteImage (Maybe Text)

The Amazon Web Services account ID associated with the registry that contains the image to delete. If you do not specify a registry, the default registry is assumed.

batchDeleteImage_repositoryName :: Lens' BatchDeleteImage Text

The repository that contains the image to delete.

batchDeleteImage_imageIds :: Lens' BatchDeleteImage [ImageIdentifier]

A list of image ID references that correspond to images to delete. The format of the imageIds reference is imageTag=tag or imageDigest=digest.

batchDeleteImageResponse_imageIds :: Lens' BatchDeleteImageResponse (Maybe [ImageIdentifier])

The image IDs of the deleted images.

batchDeleteImageResponse_failures :: Lens' BatchDeleteImageResponse (Maybe [ImageFailure])

Any failures associated with the call.

batchDeleteImageResponse_httpStatus :: Lens' BatchDeleteImageResponse Int

The response's http status code.

PutImageScanningConfiguration

putImageScanningConfiguration_registryId :: Lens' PutImageScanningConfiguration (Maybe Text)

The Amazon Web Services account ID associated with the registry that contains the repository in which to update the image scanning configuration setting. If you do not specify a registry, the default registry is assumed.

putImageScanningConfiguration_repositoryName :: Lens' PutImageScanningConfiguration Text

The name of the repository in which to update the image scanning configuration setting.

putImageScanningConfiguration_imageScanningConfiguration :: Lens' PutImageScanningConfiguration ImageScanningConfiguration

The image scanning configuration for the repository. This setting determines whether images are scanned for known vulnerabilities after being pushed to the repository.

putImageScanningConfigurationResponse_registryId :: Lens' PutImageScanningConfigurationResponse (Maybe Text)

The registry ID associated with the request.

putImageScanningConfigurationResponse_repositoryName :: Lens' PutImageScanningConfigurationResponse (Maybe Text)

The repository name associated with the request.

putImageScanningConfigurationResponse_imageScanningConfiguration :: Lens' PutImageScanningConfigurationResponse (Maybe ImageScanningConfiguration)

The image scanning configuration setting for the repository.

putImageScanningConfigurationResponse_httpStatus :: Lens' PutImageScanningConfigurationResponse Int

The response's http status code.

DeleteLifecyclePolicy

deleteLifecyclePolicy_registryId :: Lens' DeleteLifecyclePolicy (Maybe Text)

The Amazon Web Services account ID associated with the registry that contains the repository. If you do not specify a registry, the default registry is assumed.

deleteLifecyclePolicy_repositoryName :: Lens' DeleteLifecyclePolicy Text

The name of the repository.

deleteLifecyclePolicyResponse_registryId :: Lens' DeleteLifecyclePolicyResponse (Maybe Text)

The registry ID associated with the request.

deleteLifecyclePolicyResponse_repositoryName :: Lens' DeleteLifecyclePolicyResponse (Maybe Text)

The repository name associated with the request.

deleteLifecyclePolicyResponse_lifecyclePolicyText :: Lens' DeleteLifecyclePolicyResponse (Maybe Text)

The JSON lifecycle policy text.

deleteLifecyclePolicyResponse_lastEvaluatedAt :: Lens' DeleteLifecyclePolicyResponse (Maybe UTCTime)

The time stamp of the last time that the lifecycle policy was run.

deleteLifecyclePolicyResponse_httpStatus :: Lens' DeleteLifecyclePolicyResponse Int

The response's http status code.

DeleteRegistryPolicy

deleteRegistryPolicyResponse_registryId :: Lens' DeleteRegistryPolicyResponse (Maybe Text)

The registry ID associated with the request.

deleteRegistryPolicyResponse_policyText :: Lens' DeleteRegistryPolicyResponse (Maybe Text)

The contents of the registry permissions policy that was deleted.

deleteRegistryPolicyResponse_httpStatus :: Lens' DeleteRegistryPolicyResponse Int

The response's http status code.

GetRepositoryPolicy

getRepositoryPolicy_registryId :: Lens' GetRepositoryPolicy (Maybe Text)

The Amazon Web Services account ID associated with the registry that contains the repository. If you do not specify a registry, the default registry is assumed.

getRepositoryPolicy_repositoryName :: Lens' GetRepositoryPolicy Text

The name of the repository with the policy to retrieve.

getRepositoryPolicyResponse_registryId :: Lens' GetRepositoryPolicyResponse (Maybe Text)

The registry ID associated with the request.

getRepositoryPolicyResponse_repositoryName :: Lens' GetRepositoryPolicyResponse (Maybe Text)

The repository name associated with the request.

getRepositoryPolicyResponse_policyText :: Lens' GetRepositoryPolicyResponse (Maybe Text)

The JSON repository policy text associated with the repository.

getRepositoryPolicyResponse_httpStatus :: Lens' GetRepositoryPolicyResponse Int

The response's http status code.

CompleteLayerUpload

completeLayerUpload_registryId :: Lens' CompleteLayerUpload (Maybe Text)

The Amazon Web Services account ID associated with the registry to which to upload layers. If you do not specify a registry, the default registry is assumed.

completeLayerUpload_repositoryName :: Lens' CompleteLayerUpload Text

The name of the repository to associate with the image layer.

completeLayerUpload_uploadId :: Lens' CompleteLayerUpload Text

The upload ID from a previous InitiateLayerUpload operation to associate with the image layer.

completeLayerUpload_layerDigests :: Lens' CompleteLayerUpload (NonEmpty Text)

The sha256 digest of the image layer.

completeLayerUploadResponse_uploadId :: Lens' CompleteLayerUploadResponse (Maybe Text)

The upload ID associated with the layer.

completeLayerUploadResponse_registryId :: Lens' CompleteLayerUploadResponse (Maybe Text)

The registry ID associated with the request.

completeLayerUploadResponse_repositoryName :: Lens' CompleteLayerUploadResponse (Maybe Text)

The repository name associated with the request.

completeLayerUploadResponse_layerDigest :: Lens' CompleteLayerUploadResponse (Maybe Text)

The sha256 digest of the image layer.

completeLayerUploadResponse_httpStatus :: Lens' CompleteLayerUploadResponse Int

The response's http status code.

GetDownloadUrlForLayer

getDownloadUrlForLayer_registryId :: Lens' GetDownloadUrlForLayer (Maybe Text)

The Amazon Web Services account ID associated with the registry that contains the image layer to download. If you do not specify a registry, the default registry is assumed.

getDownloadUrlForLayer_repositoryName :: Lens' GetDownloadUrlForLayer Text

The name of the repository that is associated with the image layer to download.

getDownloadUrlForLayer_layerDigest :: Lens' GetDownloadUrlForLayer Text

The digest of the image layer to download.

getDownloadUrlForLayerResponse_downloadUrl :: Lens' GetDownloadUrlForLayerResponse (Maybe Text)

The pre-signed Amazon S3 download URL for the requested layer.

getDownloadUrlForLayerResponse_layerDigest :: Lens' GetDownloadUrlForLayerResponse (Maybe Text)

The digest of the image layer to download.

getDownloadUrlForLayerResponse_httpStatus :: Lens' GetDownloadUrlForLayerResponse Int

The response's http status code.

DescribeImages

describeImages_nextToken :: Lens' DescribeImages (Maybe Text)

The nextToken value returned from a previous paginated DescribeImages request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return. This option cannot be used when you specify images with imageIds.

describeImages_imageIds :: Lens' DescribeImages (Maybe [ImageIdentifier])

The list of image IDs for the requested repository.

describeImages_maxResults :: Lens' DescribeImages (Maybe Natural)

The maximum number of repository results returned by DescribeImages in paginated output. When this parameter is used, DescribeImages only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another DescribeImages request with the returned nextToken value. This value can be between 1 and 1000. If this parameter is not used, then DescribeImages returns up to 100 results and a nextToken value, if applicable. This option cannot be used when you specify images with imageIds.

describeImages_registryId :: Lens' DescribeImages (Maybe Text)

The Amazon Web Services account ID associated with the registry that contains the repository in which to describe images. If you do not specify a registry, the default registry is assumed.

describeImages_filter :: Lens' DescribeImages (Maybe DescribeImagesFilter)

The filter key and value with which to filter your DescribeImages results.

describeImages_repositoryName :: Lens' DescribeImages Text

The repository that contains the images to describe.

describeImagesResponse_nextToken :: Lens' DescribeImagesResponse (Maybe Text)

The nextToken value to include in a future DescribeImages request. When the results of a DescribeImages request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

describeImagesResponse_imageDetails :: Lens' DescribeImagesResponse (Maybe [ImageDetail])

A list of ImageDetail objects that contain data about the image.

describeImagesResponse_httpStatus :: Lens' DescribeImagesResponse Int

The response's http status code.

GetAuthorizationToken

getAuthorizationToken_registryIds :: Lens' GetAuthorizationToken (Maybe (NonEmpty Text))

A list of Amazon Web Services account IDs that are associated with the registries for which to get AuthorizationData objects. If you do not specify a registry, the default registry is assumed.

getAuthorizationTokenResponse_authorizationData :: Lens' GetAuthorizationTokenResponse (Maybe [AuthorizationData])

A list of authorization token data objects that correspond to the registryIds values in the request.

getAuthorizationTokenResponse_httpStatus :: Lens' GetAuthorizationTokenResponse Int

The response's http status code.

CreateRepository

createRepository_encryptionConfiguration :: Lens' CreateRepository (Maybe EncryptionConfiguration)

The encryption configuration for the repository. This determines how the contents of your repository are encrypted at rest.

createRepository_registryId :: Lens' CreateRepository (Maybe Text)

The AWS account ID associated with the registry to create the repository. If you do not specify a registry, the default registry is assumed.

createRepository_tags :: Lens' CreateRepository (Maybe [Tag])

The metadata that you apply to the repository to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define. Tag keys can have a maximum character length of 128 characters, and tag values can have a maximum length of 256 characters.

createRepository_imageScanningConfiguration :: Lens' CreateRepository (Maybe ImageScanningConfiguration)

The image scanning configuration for the repository. This determines whether images are scanned for known vulnerabilities after being pushed to the repository.

createRepository_imageTagMutability :: Lens' CreateRepository (Maybe ImageTagMutability)

The tag mutability setting for the repository. If this parameter is omitted, the default setting of MUTABLE will be used which will allow image tags to be overwritten. If IMMUTABLE is specified, all image tags within the repository will be immutable which will prevent them from being overwritten.

createRepository_repositoryName :: Lens' CreateRepository Text

The name to use for the repository. The repository name may be specified on its own (such as nginx-web-app) or it can be prepended with a namespace to group the repository into a category (such as project-a/nginx-web-app).

createRepositoryResponse_repository :: Lens' CreateRepositoryResponse (Maybe Repository)

The repository that was created.

createRepositoryResponse_httpStatus :: Lens' CreateRepositoryResponse Int

The response's http status code.

DeleteRepository

deleteRepository_registryId :: Lens' DeleteRepository (Maybe Text)

The Amazon Web Services account ID associated with the registry that contains the repository to delete. If you do not specify a registry, the default registry is assumed.

deleteRepository_force :: Lens' DeleteRepository (Maybe Bool)

If a repository contains images, forces the deletion.

deleteRepository_repositoryName :: Lens' DeleteRepository Text

The name of the repository to delete.

deleteRepositoryResponse_repository :: Lens' DeleteRepositoryResponse (Maybe Repository)

The repository that was deleted.

deleteRepositoryResponse_httpStatus :: Lens' DeleteRepositoryResponse Int

The response's http status code.

GetLifecyclePolicy

getLifecyclePolicy_registryId :: Lens' GetLifecyclePolicy (Maybe Text)

The Amazon Web Services account ID associated with the registry that contains the repository. If you do not specify a registry, the default registry is assumed.

getLifecyclePolicy_repositoryName :: Lens' GetLifecyclePolicy Text

The name of the repository.

getLifecyclePolicyResponse_registryId :: Lens' GetLifecyclePolicyResponse (Maybe Text)

The registry ID associated with the request.

getLifecyclePolicyResponse_repositoryName :: Lens' GetLifecyclePolicyResponse (Maybe Text)

The repository name associated with the request.

getLifecyclePolicyResponse_lifecyclePolicyText :: Lens' GetLifecyclePolicyResponse (Maybe Text)

The JSON lifecycle policy text.

getLifecyclePolicyResponse_lastEvaluatedAt :: Lens' GetLifecyclePolicyResponse (Maybe UTCTime)

The time stamp of the last time that the lifecycle policy was run.

getLifecyclePolicyResponse_httpStatus :: Lens' GetLifecyclePolicyResponse Int

The response's http status code.

StartImageScan

startImageScan_registryId :: Lens' StartImageScan (Maybe Text)

The Amazon Web Services account ID associated with the registry that contains the repository in which to start an image scan request. If you do not specify a registry, the default registry is assumed.

startImageScan_repositoryName :: Lens' StartImageScan Text

The name of the repository that contains the images to scan.

startImageScan_imageId :: Lens' StartImageScan ImageIdentifier

Undocumented member.

startImageScanResponse_imageScanStatus :: Lens' StartImageScanResponse (Maybe ImageScanStatus)

The current state of the scan.

startImageScanResponse_registryId :: Lens' StartImageScanResponse (Maybe Text)

The registry ID associated with the request.

startImageScanResponse_repositoryName :: Lens' StartImageScanResponse (Maybe Text)

The repository name associated with the request.

startImageScanResponse_imageId :: Lens' StartImageScanResponse (Maybe ImageIdentifier)

Undocumented member.

startImageScanResponse_httpStatus :: Lens' StartImageScanResponse Int

The response's http status code.

BatchCheckLayerAvailability

batchCheckLayerAvailability_registryId :: Lens' BatchCheckLayerAvailability (Maybe Text)

The Amazon Web Services account ID associated with the registry that contains the image layers to check. If you do not specify a registry, the default registry is assumed.

batchCheckLayerAvailability_repositoryName :: Lens' BatchCheckLayerAvailability Text

The name of the repository that is associated with the image layers to check.

batchCheckLayerAvailability_layerDigests :: Lens' BatchCheckLayerAvailability (NonEmpty Text)

The digests of the image layers to check.

batchCheckLayerAvailabilityResponse_failures :: Lens' BatchCheckLayerAvailabilityResponse (Maybe [LayerFailure])

Any failures associated with the call.

batchCheckLayerAvailabilityResponse_layers :: Lens' BatchCheckLayerAvailabilityResponse (Maybe [Layer])

A list of image layer objects corresponding to the image layer references in the request.

batchCheckLayerAvailabilityResponse_httpStatus :: Lens' BatchCheckLayerAvailabilityResponse Int

The response's http status code.

BatchGetImage

batchGetImage_acceptedMediaTypes :: Lens' BatchGetImage (Maybe (NonEmpty Text))

The accepted media types for the request.

Valid values: application/vnd.docker.distribution.manifest.v1+json | application/vnd.docker.distribution.manifest.v2+json | application/vnd.oci.image.manifest.v1+json

batchGetImage_registryId :: Lens' BatchGetImage (Maybe Text)

The Amazon Web Services account ID associated with the registry that contains the images to describe. If you do not specify a registry, the default registry is assumed.

batchGetImage_repositoryName :: Lens' BatchGetImage Text

The repository that contains the images to describe.

batchGetImage_imageIds :: Lens' BatchGetImage [ImageIdentifier]

A list of image ID references that correspond to images to describe. The format of the imageIds reference is imageTag=tag or imageDigest=digest.

batchGetImageResponse_images :: Lens' BatchGetImageResponse (Maybe [Image])

A list of image objects corresponding to the image references in the request.

batchGetImageResponse_failures :: Lens' BatchGetImageResponse (Maybe [ImageFailure])

Any failures associated with the call.

batchGetImageResponse_httpStatus :: Lens' BatchGetImageResponse Int

The response's http status code.

PutReplicationConfiguration

putReplicationConfiguration_replicationConfiguration :: Lens' PutReplicationConfiguration ReplicationConfiguration

An object representing the replication configuration for a registry.

putReplicationConfigurationResponse_replicationConfiguration :: Lens' PutReplicationConfigurationResponse (Maybe ReplicationConfiguration)

The contents of the replication configuration for the registry.

putReplicationConfigurationResponse_httpStatus :: Lens' PutReplicationConfigurationResponse Int

The response's http status code.

PutImageTagMutability

putImageTagMutability_registryId :: Lens' PutImageTagMutability (Maybe Text)

The Amazon Web Services account ID associated with the registry that contains the repository in which to update the image tag mutability settings. If you do not specify a registry, the default registry is assumed.

putImageTagMutability_repositoryName :: Lens' PutImageTagMutability Text

The name of the repository in which to update the image tag mutability settings.

putImageTagMutability_imageTagMutability :: Lens' PutImageTagMutability ImageTagMutability

The tag mutability setting for the repository. If MUTABLE is specified, image tags can be overwritten. If IMMUTABLE is specified, all image tags within the repository will be immutable which will prevent them from being overwritten.

putImageTagMutabilityResponse_registryId :: Lens' PutImageTagMutabilityResponse (Maybe Text)

The registry ID associated with the request.

putImageTagMutabilityResponse_repositoryName :: Lens' PutImageTagMutabilityResponse (Maybe Text)

The repository name associated with the request.

putImageTagMutabilityResponse_imageTagMutability :: Lens' PutImageTagMutabilityResponse (Maybe ImageTagMutability)

The image tag mutability setting for the repository.

putImageTagMutabilityResponse_httpStatus :: Lens' PutImageTagMutabilityResponse Int

The response's http status code.

DescribeImageReplicationStatus

describeImageReplicationStatus_registryId :: Lens' DescribeImageReplicationStatus (Maybe Text)

The Amazon Web Services account ID associated with the registry. If you do not specify a registry, the default registry is assumed.

describeImageReplicationStatus_repositoryName :: Lens' DescribeImageReplicationStatus Text

The name of the repository that the image is in.

describeImageReplicationStatus_imageId :: Lens' DescribeImageReplicationStatus ImageIdentifier

Undocumented member.

describeImageReplicationStatusResponse_repositoryName :: Lens' DescribeImageReplicationStatusResponse (Maybe Text)

The repository name associated with the request.

describeImageReplicationStatusResponse_imageId :: Lens' DescribeImageReplicationStatusResponse (Maybe ImageIdentifier)

Undocumented member.

describeImageReplicationStatusResponse_replicationStatuses :: Lens' DescribeImageReplicationStatusResponse (Maybe [ImageReplicationStatus])

The replication status details for the images in the specified repository.

describeImageReplicationStatusResponse_httpStatus :: Lens' DescribeImageReplicationStatusResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The Amazon Resource Name (ARN) that identifies the resource for which to list the tags. Currently, the only supported resource is an Amazon ECR repository.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe [Tag])

The tags for the resource.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

GetLifecyclePolicyPreview

getLifecyclePolicyPreview_nextToken :: Lens' GetLifecyclePolicyPreview (Maybe Text)

The nextToken value returned from a previous paginated GetLifecyclePolicyPreviewRequest request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return. This option cannot be used when you specify images with imageIds.

getLifecyclePolicyPreview_imageIds :: Lens' GetLifecyclePolicyPreview (Maybe [ImageIdentifier])

The list of imageIDs to be included.

getLifecyclePolicyPreview_maxResults :: Lens' GetLifecyclePolicyPreview (Maybe Natural)

The maximum number of repository results returned by GetLifecyclePolicyPreviewRequest in paginated output. When this parameter is used, GetLifecyclePolicyPreviewRequest only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another GetLifecyclePolicyPreviewRequest request with the returned nextToken value. This value can be between 1 and 1000. If this parameter is not used, then GetLifecyclePolicyPreviewRequest returns up to 100 results and a nextToken value, if applicable. This option cannot be used when you specify images with imageIds.

getLifecyclePolicyPreview_registryId :: Lens' GetLifecyclePolicyPreview (Maybe Text)

The Amazon Web Services account ID associated with the registry that contains the repository. If you do not specify a registry, the default registry is assumed.

getLifecyclePolicyPreview_filter :: Lens' GetLifecyclePolicyPreview (Maybe LifecyclePolicyPreviewFilter)

An optional parameter that filters results based on image tag status and all tags, if tagged.

getLifecyclePolicyPreview_repositoryName :: Lens' GetLifecyclePolicyPreview Text

The name of the repository.

getLifecyclePolicyPreviewResponse_nextToken :: Lens' GetLifecyclePolicyPreviewResponse (Maybe Text)

The nextToken value to include in a future GetLifecyclePolicyPreview request. When the results of a GetLifecyclePolicyPreview request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

getLifecyclePolicyPreviewResponse_status :: Lens' GetLifecyclePolicyPreviewResponse (Maybe LifecyclePolicyPreviewStatus)

The status of the lifecycle policy preview request.

getLifecyclePolicyPreviewResponse_registryId :: Lens' GetLifecyclePolicyPreviewResponse (Maybe Text)

The registry ID associated with the request.

getLifecyclePolicyPreviewResponse_repositoryName :: Lens' GetLifecyclePolicyPreviewResponse (Maybe Text)

The repository name associated with the request.

getLifecyclePolicyPreviewResponse_summary :: Lens' GetLifecyclePolicyPreviewResponse (Maybe LifecyclePolicyPreviewSummary)

The list of images that is returned as a result of the action.

getLifecyclePolicyPreviewResponse_lifecyclePolicyText :: Lens' GetLifecyclePolicyPreviewResponse (Maybe Text)

The JSON lifecycle policy text.

getLifecyclePolicyPreviewResponse_previewResults :: Lens' GetLifecyclePolicyPreviewResponse (Maybe [LifecyclePolicyPreviewResult])

The results of the lifecycle policy preview request.

getLifecyclePolicyPreviewResponse_httpStatus :: Lens' GetLifecyclePolicyPreviewResponse Int

The response's http status code.

Types

Attribute

attribute_value :: Lens' Attribute (Maybe Text)

The value assigned to the attribute key.

attribute_key :: Lens' Attribute Text

The attribute key.

AuthorizationData

authorizationData_proxyEndpoint :: Lens' AuthorizationData (Maybe Text)

The registry URL to use for this authorization token in a docker login command. The Amazon ECR registry URL format is https://aws_account_id.dkr.ecr.region.amazonaws.com. For example, https://012345678910.dkr.ecr.us-east-1.amazonaws.com..

authorizationData_expiresAt :: Lens' AuthorizationData (Maybe UTCTime)

The Unix time in seconds and milliseconds when the authorization token expires. Authorization tokens are valid for 12 hours.

authorizationData_authorizationToken :: Lens' AuthorizationData (Maybe Text)

A base64-encoded string that contains authorization data for the specified Amazon ECR registry. When the string is decoded, it is presented in the format user:password for private registry authentication using docker login.

DescribeImagesFilter

describeImagesFilter_tagStatus :: Lens' DescribeImagesFilter (Maybe TagStatus)

The tag status with which to filter your DescribeImages results. You can filter results based on whether they are TAGGED or UNTAGGED.

EncryptionConfiguration

encryptionConfiguration_kmsKey :: Lens' EncryptionConfiguration (Maybe Text)

If you use the KMS encryption type, specify the KMS key to use for encryption. The alias, key ID, or full ARN of the KMS key can be specified. The key must exist in the same Region as the repository. If no key is specified, the default Amazon Web Services managed KMS key for Amazon ECR will be used.

encryptionConfiguration_encryptionType :: Lens' EncryptionConfiguration EncryptionType

The encryption type to use.

If you use the KMS encryption type, the contents of the repository will be encrypted using server-side encryption with Key Management Service key stored in KMS. When you use KMS to encrypt your data, you can either use the default Amazon Web Services managed KMS key for Amazon ECR, or specify your own KMS key, which you already created. For more information, see Protecting data using server-side encryption with an KMS key stored in Key Management Service (SSE-KMS) in the Amazon Simple Storage Service Console Developer Guide..

If you use the AES256 encryption type, Amazon ECR uses server-side encryption with Amazon S3-managed encryption keys which encrypts the images in the repository using an AES-256 encryption algorithm. For more information, see Protecting data using server-side encryption with Amazon S3-managed encryption keys (SSE-S3) in the Amazon Simple Storage Service Console Developer Guide..

Image

image_registryId :: Lens' Image (Maybe Text)

The Amazon Web Services account ID associated with the registry containing the image.

image_imageManifest :: Lens' Image (Maybe Text)

The image manifest associated with the image.

image_repositoryName :: Lens' Image (Maybe Text)

The name of the repository associated with the image.

image_imageId :: Lens' Image (Maybe ImageIdentifier)

An object containing the image tag and image digest associated with an image.

image_imageManifestMediaType :: Lens' Image (Maybe Text)

The manifest media type of the image.

ImageDetail

imageDetail_imageDigest :: Lens' ImageDetail (Maybe Text)

The sha256 digest of the image manifest.

imageDetail_imageScanStatus :: Lens' ImageDetail (Maybe ImageScanStatus)

The current state of the scan.

imageDetail_imageTags :: Lens' ImageDetail (Maybe [Text])

The list of tags associated with this image.

imageDetail_registryId :: Lens' ImageDetail (Maybe Text)

The Amazon Web Services account ID associated with the registry to which this image belongs.

imageDetail_repositoryName :: Lens' ImageDetail (Maybe Text)

The name of the repository to which this image belongs.

imageDetail_artifactMediaType :: Lens' ImageDetail (Maybe Text)

The artifact media type of the image.

imageDetail_imageSizeInBytes :: Lens' ImageDetail (Maybe Integer)

The size, in bytes, of the image in the repository.

If the image is a manifest list, this will be the max size of all manifests in the list.

Beginning with Docker version 1.9, the Docker client compresses image layers before pushing them to a V2 Docker registry. The output of the docker images command shows the uncompressed image size, so it may return a larger image size than the image sizes returned by DescribeImages.

imageDetail_imageManifestMediaType :: Lens' ImageDetail (Maybe Text)

The media type of the image manifest.

imageDetail_imagePushedAt :: Lens' ImageDetail (Maybe UTCTime)

The date and time, expressed in standard JavaScript date format, at which the current image was pushed to the repository.

imageDetail_imageScanFindingsSummary :: Lens' ImageDetail (Maybe ImageScanFindingsSummary)

A summary of the last completed image scan.

ImageFailure

imageFailure_failureCode :: Lens' ImageFailure (Maybe ImageFailureCode)

The code associated with the failure.

imageFailure_imageId :: Lens' ImageFailure (Maybe ImageIdentifier)

The image ID associated with the failure.

imageFailure_failureReason :: Lens' ImageFailure (Maybe Text)

The reason for the failure.

ImageIdentifier

imageIdentifier_imageDigest :: Lens' ImageIdentifier (Maybe Text)

The sha256 digest of the image manifest.

imageIdentifier_imageTag :: Lens' ImageIdentifier (Maybe Text)

The tag used for the image.

ImageReplicationStatus

imageReplicationStatus_status :: Lens' ImageReplicationStatus (Maybe ReplicationStatus)

The image replication status.

imageReplicationStatus_registryId :: Lens' ImageReplicationStatus (Maybe Text)

The AWS account ID associated with the registry to which the image belongs.

imageReplicationStatus_failureCode :: Lens' ImageReplicationStatus (Maybe Text)

The failure code for a replication that has failed.

imageReplicationStatus_region :: Lens' ImageReplicationStatus (Maybe Text)

The destination Region for the image replication.

ImageScanFinding

imageScanFinding_uri :: Lens' ImageScanFinding (Maybe Text)

A link containing additional details about the security vulnerability.

imageScanFinding_severity :: Lens' ImageScanFinding (Maybe FindingSeverity)

The finding severity.

imageScanFinding_name :: Lens' ImageScanFinding (Maybe Text)

The name associated with the finding, usually a CVE number.

imageScanFinding_attributes :: Lens' ImageScanFinding (Maybe [Attribute])

A collection of attributes of the host from which the finding is generated.

imageScanFinding_description :: Lens' ImageScanFinding (Maybe Text)

The description of the finding.

ImageScanFindings

imageScanFindings_findings :: Lens' ImageScanFindings (Maybe [ImageScanFinding])

The findings from the image scan.

imageScanFindings_imageScanCompletedAt :: Lens' ImageScanFindings (Maybe UTCTime)

The time of the last completed image scan.

imageScanFindings_vulnerabilitySourceUpdatedAt :: Lens' ImageScanFindings (Maybe UTCTime)

The time when the vulnerability data was last scanned.

imageScanFindings_findingSeverityCounts :: Lens' ImageScanFindings (Maybe (HashMap FindingSeverity Natural))

The image vulnerability counts, sorted by severity.

ImageScanFindingsSummary

imageScanFindingsSummary_imageScanCompletedAt :: Lens' ImageScanFindingsSummary (Maybe UTCTime)

The time of the last completed image scan.

imageScanFindingsSummary_vulnerabilitySourceUpdatedAt :: Lens' ImageScanFindingsSummary (Maybe UTCTime)

The time when the vulnerability data was last scanned.

imageScanFindingsSummary_findingSeverityCounts :: Lens' ImageScanFindingsSummary (Maybe (HashMap FindingSeverity Natural))

The image vulnerability counts, sorted by severity.

ImageScanStatus

imageScanStatus_status :: Lens' ImageScanStatus (Maybe ScanStatus)

The current state of an image scan.

imageScanStatus_description :: Lens' ImageScanStatus (Maybe Text)

The description of the image scan status.

ImageScanningConfiguration

imageScanningConfiguration_scanOnPush :: Lens' ImageScanningConfiguration (Maybe Bool)

The setting that determines whether images are scanned after being pushed to a repository. If set to true, images will be scanned after being pushed. If this parameter is not specified, it will default to false and images will not be scanned unless a scan is manually started with the API_StartImageScan API.

Layer

layer_layerSize :: Lens' Layer (Maybe Integer)

The size, in bytes, of the image layer.

layer_layerAvailability :: Lens' Layer (Maybe LayerAvailability)

The availability status of the image layer.

layer_mediaType :: Lens' Layer (Maybe Text)

The media type of the layer, such as application/vnd.docker.image.rootfs.diff.tar.gzip or application/vnd.oci.image.layer.v1.tar+gzip.

layer_layerDigest :: Lens' Layer (Maybe Text)

The sha256 digest of the image layer.

LayerFailure

layerFailure_failureCode :: Lens' LayerFailure (Maybe LayerFailureCode)

The failure code associated with the failure.

layerFailure_failureReason :: Lens' LayerFailure (Maybe Text)

The reason for the failure.

layerFailure_layerDigest :: Lens' LayerFailure (Maybe Text)

The layer digest associated with the failure.

LifecyclePolicyPreviewFilter

lifecyclePolicyPreviewFilter_tagStatus :: Lens' LifecyclePolicyPreviewFilter (Maybe TagStatus)

The tag status of the image.

LifecyclePolicyPreviewResult

lifecyclePolicyPreviewResult_imageDigest :: Lens' LifecyclePolicyPreviewResult (Maybe Text)

The sha256 digest of the image manifest.

lifecyclePolicyPreviewResult_appliedRulePriority :: Lens' LifecyclePolicyPreviewResult (Maybe Natural)

The priority of the applied rule.

lifecyclePolicyPreviewResult_imageTags :: Lens' LifecyclePolicyPreviewResult (Maybe [Text])

The list of tags associated with this image.

lifecyclePolicyPreviewResult_action :: Lens' LifecyclePolicyPreviewResult (Maybe LifecyclePolicyRuleAction)

The type of action to be taken.

lifecyclePolicyPreviewResult_imagePushedAt :: Lens' LifecyclePolicyPreviewResult (Maybe UTCTime)

The date and time, expressed in standard JavaScript date format, at which the current image was pushed to the repository.

LifecyclePolicyPreviewSummary

lifecyclePolicyPreviewSummary_expiringImageTotalCount :: Lens' LifecyclePolicyPreviewSummary (Maybe Natural)

The number of expiring images.

LifecyclePolicyRuleAction

lifecyclePolicyRuleAction_type :: Lens' LifecyclePolicyRuleAction (Maybe ImageActionType)

The type of action to be taken.

ListImagesFilter

listImagesFilter_tagStatus :: Lens' ListImagesFilter (Maybe TagStatus)

The tag status with which to filter your ListImages results. You can filter results based on whether they are TAGGED or UNTAGGED.

ReplicationConfiguration

replicationConfiguration_rules :: Lens' ReplicationConfiguration [ReplicationRule]

An array of objects representing the replication destinations and repository filters for a replication configuration.

ReplicationDestination

replicationDestination_region :: Lens' ReplicationDestination Text

The Region to replicate to.

replicationDestination_registryId :: Lens' ReplicationDestination Text

The Amazon Web Services account ID of the Amazon ECR private registry to replicate to. When configuring cross-Region replication within your own registry, specify your own account ID.

ReplicationRule

replicationRule_repositoryFilters :: Lens' ReplicationRule (Maybe (NonEmpty RepositoryFilter))

An array of objects representing the filters for a replication rule. Specifying a repository filter for a replication rule provides a method for controlling which repositories in a private registry are replicated.

replicationRule_destinations :: Lens' ReplicationRule [ReplicationDestination]

An array of objects representing the destination for a replication rule.

Repository

repository_encryptionConfiguration :: Lens' Repository (Maybe EncryptionConfiguration)

The encryption configuration for the repository. This determines how the contents of your repository are encrypted at rest.

repository_repositoryUri :: Lens' Repository (Maybe Text)

The URI for the repository. You can use this URI for container image push and pull operations.

repository_registryId :: Lens' Repository (Maybe Text)

The Amazon Web Services account ID associated with the registry that contains the repository.

repository_createdAt :: Lens' Repository (Maybe UTCTime)

The date and time, in JavaScript date format, when the repository was created.

repository_repositoryName :: Lens' Repository (Maybe Text)

The name of the repository.

repository_repositoryArn :: Lens' Repository (Maybe Text)

The Amazon Resource Name (ARN) that identifies the repository. The ARN contains the arn:aws:ecr namespace, followed by the region of the repository, Amazon Web Services account ID of the repository owner, repository namespace, and repository name. For example, arn:aws:ecr:region:012345678910:repository/test.

repository_imageScanningConfiguration :: Lens' Repository (Maybe ImageScanningConfiguration)

Undocumented member.

repository_imageTagMutability :: Lens' Repository (Maybe ImageTagMutability)

The tag mutability setting for the repository.

RepositoryFilter

repositoryFilter_filter :: Lens' RepositoryFilter Text

The repository filter details. When the PREFIX_MATCH filter type is specified, this value is required and should be the repository name prefix to configure replication for.

repositoryFilter_filterType :: Lens' RepositoryFilter RepositoryFilterType

The repository filter type. The only supported value is PREFIX_MATCH, which is a repository name prefix specified with the filter parameter.

Tag

tag_key :: Lens' Tag (Maybe Text)

One part of a key-value pair that make up a tag. A key is a general label that acts like a category for more specific tag values.

tag_value :: Lens' Tag (Maybe Text)

The optional part of a key-value pair that make up a tag. A value acts as a descriptor within a tag category (key).