Network.AWS.ECR.PutLifecyclePolicy

Description

Creates or updates the lifecycle policy for the specified repository. For more information, see Lifecycle policy template.

Creating a Request

data PutLifecyclePolicy

See: newPutLifecyclePolicy smart constructor.

Constructors

PutLifecyclePolicy'

Fields

registryId :: Maybe Text

The Amazon Web Services account ID associated with the registry that contains the repository. If you do not specify a registry, the default registry is assumed.

repositoryName :: Text

The name of the repository to receive the policy.

lifecyclePolicyText :: Text

The JSON repository policy text to apply to the repository.

newPutLifecyclePolicy

Arguments

:: Text
$sel:repositoryName:PutLifecyclePolicy'

-> Text
$sel:lifecyclePolicyText:PutLifecyclePolicy'

-> PutLifecyclePolicy

Create a value of PutLifecyclePolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:registryId:PutLifecyclePolicy', putLifecyclePolicy_registryId - The Amazon Web Services account ID associated with the registry that contains the repository. If you do not specify a registry, the default registry is assumed.

$sel:repositoryName:PutLifecyclePolicy', putLifecyclePolicy_repositoryName - The name of the repository to receive the policy.

$sel:lifecyclePolicyText:PutLifecyclePolicy', putLifecyclePolicy_lifecyclePolicyText - The JSON repository policy text to apply to the repository.

Request Lenses

putLifecyclePolicy_registryId :: Lens' PutLifecyclePolicy (Maybe Text)

The Amazon Web Services account ID associated with the registry that contains the repository. If you do not specify a registry, the default registry is assumed.

putLifecyclePolicy_repositoryName :: Lens' PutLifecyclePolicy Text

The name of the repository to receive the policy.

putLifecyclePolicy_lifecyclePolicyText :: Lens' PutLifecyclePolicy Text

The JSON repository policy text to apply to the repository.

Destructuring the Response

data PutLifecyclePolicyResponse

See: newPutLifecyclePolicyResponse smart constructor.

Constructors

PutLifecyclePolicyResponse'

Fields

registryId :: Maybe Text

The registry ID associated with the request.

repositoryName :: Maybe Text

The repository name associated with the request.

lifecyclePolicyText :: Maybe Text

The JSON repository policy text.

httpStatus :: Int

The response's http status code.

newPutLifecyclePolicyResponse

Arguments

:: Int
$sel:httpStatus:PutLifecyclePolicyResponse'

-> PutLifecyclePolicyResponse

Create a value of PutLifecyclePolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:registryId:PutLifecyclePolicy', putLifecyclePolicyResponse_registryId - The registry ID associated with the request.

$sel:repositoryName:PutLifecyclePolicy', putLifecyclePolicyResponse_repositoryName - The repository name associated with the request.

$sel:lifecyclePolicyText:PutLifecyclePolicy', putLifecyclePolicyResponse_lifecyclePolicyText - The JSON repository policy text.

$sel:httpStatus:PutLifecyclePolicyResponse', putLifecyclePolicyResponse_httpStatus - The response's http status code.

Response Lenses

putLifecyclePolicyResponse_registryId :: Lens' PutLifecyclePolicyResponse (Maybe Text)

The registry ID associated with the request.

putLifecyclePolicyResponse_repositoryName :: Lens' PutLifecyclePolicyResponse (Maybe Text)

The repository name associated with the request.

putLifecyclePolicyResponse_lifecyclePolicyText :: Lens' PutLifecyclePolicyResponse (Maybe Text)

The JSON repository policy text.

putLifecyclePolicyResponse_httpStatus :: Lens' PutLifecyclePolicyResponse Int

The response's http status code.