Network.AWS.ECR.StartLifecyclePolicyPreview

Description

Starts a preview of a lifecycle policy for the specified repository. This allows you to see the results before associating the lifecycle policy with the repository.

Creating a Request

data StartLifecyclePolicyPreview

See: newStartLifecyclePolicyPreview smart constructor.

Constructors

StartLifecyclePolicyPreview'

Fields

registryId :: Maybe Text

The Amazon Web Services account ID associated with the registry that contains the repository. If you do not specify a registry, the default registry is assumed.

lifecyclePolicyText :: Maybe Text

The policy to be evaluated against. If you do not specify a policy, the current policy for the repository is used.

repositoryName :: Text

The name of the repository to be evaluated.

newStartLifecyclePolicyPreview

Arguments

:: Text
$sel:repositoryName:StartLifecyclePolicyPreview'

-> StartLifecyclePolicyPreview

Create a value of StartLifecyclePolicyPreview with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:registryId:StartLifecyclePolicyPreview', startLifecyclePolicyPreview_registryId - The Amazon Web Services account ID associated with the registry that contains the repository. If you do not specify a registry, the default registry is assumed.

$sel:lifecyclePolicyText:StartLifecyclePolicyPreview', startLifecyclePolicyPreview_lifecyclePolicyText - The policy to be evaluated against. If you do not specify a policy, the current policy for the repository is used.

$sel:repositoryName:StartLifecyclePolicyPreview', startLifecyclePolicyPreview_repositoryName - The name of the repository to be evaluated.

Request Lenses

startLifecyclePolicyPreview_registryId :: Lens' StartLifecyclePolicyPreview (Maybe Text)

The Amazon Web Services account ID associated with the registry that contains the repository. If you do not specify a registry, the default registry is assumed.

startLifecyclePolicyPreview_lifecyclePolicyText :: Lens' StartLifecyclePolicyPreview (Maybe Text)

The policy to be evaluated against. If you do not specify a policy, the current policy for the repository is used.

startLifecyclePolicyPreview_repositoryName :: Lens' StartLifecyclePolicyPreview Text

The name of the repository to be evaluated.

Destructuring the Response

data StartLifecyclePolicyPreviewResponse

See: newStartLifecyclePolicyPreviewResponse smart constructor.

Constructors

StartLifecyclePolicyPreviewResponse'

Fields

status :: Maybe LifecyclePolicyPreviewStatus

The status of the lifecycle policy preview request.

registryId :: Maybe Text

The registry ID associated with the request.

lifecyclePolicyText :: Maybe Text

The JSON repository policy text.

repositoryName :: Maybe Text

The repository name associated with the request.

httpStatus :: Int

The response's http status code.

newStartLifecyclePolicyPreviewResponse

Arguments

:: Int
$sel:httpStatus:StartLifecyclePolicyPreviewResponse'

-> StartLifecyclePolicyPreviewResponse

Create a value of StartLifecyclePolicyPreviewResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:StartLifecyclePolicyPreviewResponse', startLifecyclePolicyPreviewResponse_status - The status of the lifecycle policy preview request.

$sel:registryId:StartLifecyclePolicyPreview', startLifecyclePolicyPreviewResponse_registryId - The registry ID associated with the request.

$sel:lifecyclePolicyText:StartLifecyclePolicyPreview', startLifecyclePolicyPreviewResponse_lifecyclePolicyText - The JSON repository policy text.

$sel:repositoryName:StartLifecyclePolicyPreview', startLifecyclePolicyPreviewResponse_repositoryName - The repository name associated with the request.

$sel:httpStatus:StartLifecyclePolicyPreviewResponse', startLifecyclePolicyPreviewResponse_httpStatus - The response's http status code.

Response Lenses

startLifecyclePolicyPreviewResponse_status :: Lens' StartLifecyclePolicyPreviewResponse (Maybe LifecyclePolicyPreviewStatus)

The status of the lifecycle policy preview request.

startLifecyclePolicyPreviewResponse_registryId :: Lens' StartLifecyclePolicyPreviewResponse (Maybe Text)

The registry ID associated with the request.

startLifecyclePolicyPreviewResponse_lifecyclePolicyText :: Lens' StartLifecyclePolicyPreviewResponse (Maybe Text)

The JSON repository policy text.

startLifecyclePolicyPreviewResponse_repositoryName :: Lens' StartLifecyclePolicyPreviewResponse (Maybe Text)

The repository name associated with the request.

startLifecyclePolicyPreviewResponse_httpStatus :: Lens' StartLifecyclePolicyPreviewResponse Int

The response's http status code.