Network.AWS.ECR.Types.ImageScanFindings

Description

Documentation

data ImageScanFindings

The details of an image scan.

See: newImageScanFindings smart constructor.

Constructors

ImageScanFindings'

Fields

findings :: Maybe [ImageScanFinding]

The findings from the image scan.

imageScanCompletedAt :: Maybe POSIX

The time of the last completed image scan.

vulnerabilitySourceUpdatedAt :: Maybe POSIX

The time when the vulnerability data was last scanned.

findingSeverityCounts :: Maybe (HashMap FindingSeverity Natural)

The image vulnerability counts, sorted by severity.

newImageScanFindings :: ImageScanFindings

Create a value of ImageScanFindings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:findings:ImageScanFindings', imageScanFindings_findings - The findings from the image scan.

$sel:imageScanCompletedAt:ImageScanFindings', imageScanFindings_imageScanCompletedAt - The time of the last completed image scan.

$sel:vulnerabilitySourceUpdatedAt:ImageScanFindings', imageScanFindings_vulnerabilitySourceUpdatedAt - The time when the vulnerability data was last scanned.

$sel:findingSeverityCounts:ImageScanFindings', imageScanFindings_findingSeverityCounts - The image vulnerability counts, sorted by severity.

imageScanFindings_findings :: Lens' ImageScanFindings (Maybe [ImageScanFinding])

The findings from the image scan.

imageScanFindings_imageScanCompletedAt :: Lens' ImageScanFindings (Maybe UTCTime)

The time of the last completed image scan.

imageScanFindings_vulnerabilitySourceUpdatedAt :: Lens' ImageScanFindings (Maybe UTCTime)

The time when the vulnerability data was last scanned.

imageScanFindings_findingSeverityCounts :: Lens' ImageScanFindings (Maybe (HashMap FindingSeverity Natural))

The image vulnerability counts, sorted by severity.