Network.AWS.ECR.Types.ImageScanFindingsSummary

Description

Documentation

data ImageScanFindingsSummary

A summary of the last completed image scan.

See: newImageScanFindingsSummary smart constructor.

Constructors

ImageScanFindingsSummary'

Fields

imageScanCompletedAt :: Maybe POSIX

The time of the last completed image scan.

vulnerabilitySourceUpdatedAt :: Maybe POSIX

The time when the vulnerability data was last scanned.

findingSeverityCounts :: Maybe (HashMap FindingSeverity Natural)

The image vulnerability counts, sorted by severity.

newImageScanFindingsSummary :: ImageScanFindingsSummary

Create a value of ImageScanFindingsSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:imageScanCompletedAt:ImageScanFindingsSummary', imageScanFindingsSummary_imageScanCompletedAt - The time of the last completed image scan.

$sel:vulnerabilitySourceUpdatedAt:ImageScanFindingsSummary', imageScanFindingsSummary_vulnerabilitySourceUpdatedAt - The time when the vulnerability data was last scanned.

$sel:findingSeverityCounts:ImageScanFindingsSummary', imageScanFindingsSummary_findingSeverityCounts - The image vulnerability counts, sorted by severity.

imageScanFindingsSummary_imageScanCompletedAt :: Lens' ImageScanFindingsSummary (Maybe UTCTime)

The time of the last completed image scan.

imageScanFindingsSummary_vulnerabilitySourceUpdatedAt :: Lens' ImageScanFindingsSummary (Maybe UTCTime)

The time when the vulnerability data was last scanned.

imageScanFindingsSummary_findingSeverityCounts :: Lens' ImageScanFindingsSummary (Maybe (HashMap FindingSeverity Natural))

The image vulnerability counts, sorted by severity.