Network.AWS.ECR.UploadLayerPart

Description

Uploads an image layer part to Amazon ECR.

When an image is pushed, each new image layer is uploaded in parts. The maximum size of each image layer part can be 20971520 bytes (or about 20MB). The UploadLayerPart API is called once per each new image layer part.

This operation is used by the Amazon ECR proxy and is not generally used by customers for pulling and pushing images. In most cases, you should use the docker CLI to pull, tag, and push images.

Creating a Request

data UploadLayerPart

See: newUploadLayerPart smart constructor.

Constructors

UploadLayerPart'

Fields

registryId :: Maybe Text

The Amazon Web Services account ID associated with the registry to which you are uploading layer parts. If you do not specify a registry, the default registry is assumed.

repositoryName :: Text

The name of the repository to which you are uploading layer parts.

uploadId :: Text

The upload ID from a previous InitiateLayerUpload operation to associate with the layer part upload.

partFirstByte :: Natural

The position of the first byte of the layer part witin the overall image layer.

partLastByte :: Natural

The position of the last byte of the layer part within the overall image layer.

layerPartBlob :: Base64

The base64-encoded layer part payload.

newUploadLayerPart

Arguments

:: Text
$sel:repositoryName:UploadLayerPart'

-> Text
$sel:uploadId:UploadLayerPart'

-> Natural
$sel:partFirstByte:UploadLayerPart'

-> Natural
$sel:partLastByte:UploadLayerPart'

-> ByteString
$sel:layerPartBlob:UploadLayerPart'

-> UploadLayerPart

Create a value of UploadLayerPart with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:registryId:UploadLayerPart', uploadLayerPart_registryId - The Amazon Web Services account ID associated with the registry to which you are uploading layer parts. If you do not specify a registry, the default registry is assumed.

$sel:repositoryName:UploadLayerPart', uploadLayerPart_repositoryName - The name of the repository to which you are uploading layer parts.

$sel:uploadId:UploadLayerPart', uploadLayerPart_uploadId - The upload ID from a previous InitiateLayerUpload operation to associate with the layer part upload.

$sel:partFirstByte:UploadLayerPart', uploadLayerPart_partFirstByte - The position of the first byte of the layer part witin the overall image layer.

$sel:partLastByte:UploadLayerPart', uploadLayerPart_partLastByte - The position of the last byte of the layer part within the overall image layer.

$sel:layerPartBlob:UploadLayerPart', uploadLayerPart_layerPartBlob - The base64-encoded layer part payload.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

Request Lenses

uploadLayerPart_registryId :: Lens' UploadLayerPart (Maybe Text)

The Amazon Web Services account ID associated with the registry to which you are uploading layer parts. If you do not specify a registry, the default registry is assumed.

uploadLayerPart_repositoryName :: Lens' UploadLayerPart Text

The name of the repository to which you are uploading layer parts.

uploadLayerPart_uploadId :: Lens' UploadLayerPart Text

The upload ID from a previous InitiateLayerUpload operation to associate with the layer part upload.

uploadLayerPart_partFirstByte :: Lens' UploadLayerPart Natural

The position of the first byte of the layer part witin the overall image layer.

uploadLayerPart_partLastByte :: Lens' UploadLayerPart Natural

The position of the last byte of the layer part within the overall image layer.

uploadLayerPart_layerPartBlob :: Lens' UploadLayerPart ByteString

The base64-encoded layer part payload.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

Destructuring the Response

data UploadLayerPartResponse

See: newUploadLayerPartResponse smart constructor.

Constructors

UploadLayerPartResponse'

Fields

uploadId :: Maybe Text

The upload ID associated with the request.

registryId :: Maybe Text

The registry ID associated with the request.

repositoryName :: Maybe Text

The repository name associated with the request.

lastByteReceived :: Maybe Natural

The integer value of the last byte received in the request.

httpStatus :: Int

The response's http status code.

newUploadLayerPartResponse

Arguments

:: Int
$sel:httpStatus:UploadLayerPartResponse'

-> UploadLayerPartResponse

Create a value of UploadLayerPartResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:uploadId:UploadLayerPart', uploadLayerPartResponse_uploadId - The upload ID associated with the request.

$sel:registryId:UploadLayerPart', uploadLayerPartResponse_registryId - The registry ID associated with the request.

$sel:repositoryName:UploadLayerPart', uploadLayerPartResponse_repositoryName - The repository name associated with the request.

$sel:lastByteReceived:UploadLayerPartResponse', uploadLayerPartResponse_lastByteReceived - The integer value of the last byte received in the request.

$sel:httpStatus:UploadLayerPartResponse', uploadLayerPartResponse_httpStatus - The response's http status code.

Response Lenses

uploadLayerPartResponse_uploadId :: Lens' UploadLayerPartResponse (Maybe Text)

The upload ID associated with the request.

uploadLayerPartResponse_registryId :: Lens' UploadLayerPartResponse (Maybe Text)

The registry ID associated with the request.

uploadLayerPartResponse_repositoryName :: Lens' UploadLayerPartResponse (Maybe Text)

The repository name associated with the request.

uploadLayerPartResponse_lastByteReceived :: Lens' UploadLayerPartResponse (Maybe Natural)

The integer value of the last byte received in the request.

uploadLayerPartResponse_httpStatus :: Lens' UploadLayerPartResponse Int

The response's http status code.