Network.AWS.ECRPublic.CompleteLayerUpload

Description

Informs Amazon ECR that the image layer upload has completed for a specified public registry, repository name, and upload ID. You can optionally provide a sha256 digest of the image layer for data validation purposes.

When an image is pushed, the CompleteLayerUpload API is called once per each new image layer to verify that the upload has completed.

This operation is used by the Amazon ECR proxy and is not generally used by customers for pulling and pushing images. In most cases, you should use the docker CLI to pull, tag, and push images.

Creating a Request

data CompleteLayerUpload

See: newCompleteLayerUpload smart constructor.

Constructors

CompleteLayerUpload'

Fields

registryId :: Maybe Text

The AWS account ID associated with the registry to which to upload layers. If you do not specify a registry, the default public registry is assumed.

repositoryName :: Text

The name of the repository in a public registry to associate with the image layer.

uploadId :: Text

The upload ID from a previous InitiateLayerUpload operation to associate with the image layer.

layerDigests :: NonEmpty Text

The sha256 digest of the image layer.

newCompleteLayerUpload

Arguments

:: Text
$sel:repositoryName:CompleteLayerUpload'

-> Text
$sel:uploadId:CompleteLayerUpload'

-> NonEmpty Text
$sel:layerDigests:CompleteLayerUpload'

-> CompleteLayerUpload

Create a value of CompleteLayerUpload with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:registryId:CompleteLayerUpload', completeLayerUpload_registryId - The AWS account ID associated with the registry to which to upload layers. If you do not specify a registry, the default public registry is assumed.

$sel:repositoryName:CompleteLayerUpload', completeLayerUpload_repositoryName - The name of the repository in a public registry to associate with the image layer.

$sel:uploadId:CompleteLayerUpload', completeLayerUpload_uploadId - The upload ID from a previous InitiateLayerUpload operation to associate with the image layer.

$sel:layerDigests:CompleteLayerUpload', completeLayerUpload_layerDigests - The sha256 digest of the image layer.

Request Lenses

completeLayerUpload_registryId :: Lens' CompleteLayerUpload (Maybe Text)

The AWS account ID associated with the registry to which to upload layers. If you do not specify a registry, the default public registry is assumed.

completeLayerUpload_repositoryName :: Lens' CompleteLayerUpload Text

The name of the repository in a public registry to associate with the image layer.

completeLayerUpload_uploadId :: Lens' CompleteLayerUpload Text

The upload ID from a previous InitiateLayerUpload operation to associate with the image layer.

completeLayerUpload_layerDigests :: Lens' CompleteLayerUpload (NonEmpty Text)

The sha256 digest of the image layer.

Destructuring the Response

data CompleteLayerUploadResponse

See: newCompleteLayerUploadResponse smart constructor.

Constructors

CompleteLayerUploadResponse'

Fields

registryId :: Maybe Text

The public registry ID associated with the request.

layerDigest :: Maybe Text

The sha256 digest of the image layer.

repositoryName :: Maybe Text

The repository name associated with the request.

uploadId :: Maybe Text

The upload ID associated with the layer.

httpStatus :: Int

The response's http status code.

newCompleteLayerUploadResponse

Arguments

:: Int
$sel:httpStatus:CompleteLayerUploadResponse'

-> CompleteLayerUploadResponse

Create a value of CompleteLayerUploadResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:registryId:CompleteLayerUpload', completeLayerUploadResponse_registryId - The public registry ID associated with the request.

$sel:layerDigest:CompleteLayerUploadResponse', completeLayerUploadResponse_layerDigest - The sha256 digest of the image layer.

$sel:repositoryName:CompleteLayerUpload', completeLayerUploadResponse_repositoryName - The repository name associated with the request.

$sel:uploadId:CompleteLayerUpload', completeLayerUploadResponse_uploadId - The upload ID associated with the layer.

$sel:httpStatus:CompleteLayerUploadResponse', completeLayerUploadResponse_httpStatus - The response's http status code.

Response Lenses

completeLayerUploadResponse_registryId :: Lens' CompleteLayerUploadResponse (Maybe Text)

The public registry ID associated with the request.

completeLayerUploadResponse_layerDigest :: Lens' CompleteLayerUploadResponse (Maybe Text)

The sha256 digest of the image layer.

completeLayerUploadResponse_repositoryName :: Lens' CompleteLayerUploadResponse (Maybe Text)

The repository name associated with the request.

completeLayerUploadResponse_uploadId :: Lens' CompleteLayerUploadResponse (Maybe Text)

The upload ID associated with the layer.

completeLayerUploadResponse_httpStatus :: Lens' CompleteLayerUploadResponse Int

The response's http status code.