Network.AWS.ECRPublic.DescribeRegistries

Description

Returns details for a public registry.

This operation returns paginated results.

Creating a Request

data DescribeRegistries

See: newDescribeRegistries smart constructor.

Constructors

DescribeRegistries'

Fields

nextToken :: Maybe Text

The nextToken value returned from a previous paginated DescribeRegistries request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return.

This token should be treated as an opaque identifier that is only used to retrieve the next items in a list and not for other programmatic purposes.

maxResults :: Maybe Natural

The maximum number of repository results returned by DescribeRegistries in paginated output. When this parameter is used, DescribeRegistries only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another DescribeRegistries request with the returned nextToken value. This value can be between 1 and 1000. If this parameter is not used, then DescribeRegistries returns up to 100 results and a nextToken value, if applicable.

newDescribeRegistries :: DescribeRegistries

Create a value of DescribeRegistries with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeRegistries', describeRegistries_nextToken - The nextToken value returned from a previous paginated DescribeRegistries request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return.

This token should be treated as an opaque identifier that is only used to retrieve the next items in a list and not for other programmatic purposes.

$sel:maxResults:DescribeRegistries', describeRegistries_maxResults - The maximum number of repository results returned by DescribeRegistries in paginated output. When this parameter is used, DescribeRegistries only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another DescribeRegistries request with the returned nextToken value. This value can be between 1 and 1000. If this parameter is not used, then DescribeRegistries returns up to 100 results and a nextToken value, if applicable.

Request Lenses

describeRegistries_nextToken :: Lens' DescribeRegistries (Maybe Text)

The nextToken value returned from a previous paginated DescribeRegistries request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return.

This token should be treated as an opaque identifier that is only used to retrieve the next items in a list and not for other programmatic purposes.

describeRegistries_maxResults :: Lens' DescribeRegistries (Maybe Natural)

The maximum number of repository results returned by DescribeRegistries in paginated output. When this parameter is used, DescribeRegistries only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another DescribeRegistries request with the returned nextToken value. This value can be between 1 and 1000. If this parameter is not used, then DescribeRegistries returns up to 100 results and a nextToken value, if applicable.

Destructuring the Response

data DescribeRegistriesResponse

See: newDescribeRegistriesResponse smart constructor.

Constructors

DescribeRegistriesResponse'

Fields

nextToken :: Maybe Text

The nextToken value to include in a future DescribeRepositories request. When the results of a DescribeRepositories request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

httpStatus :: Int

The response's http status code.

registries :: [Registry]

An object containing the details for a public registry.

newDescribeRegistriesResponse

Arguments

:: Int
$sel:httpStatus:DescribeRegistriesResponse'

-> DescribeRegistriesResponse

Create a value of DescribeRegistriesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeRegistries', describeRegistriesResponse_nextToken - The nextToken value to include in a future DescribeRepositories request. When the results of a DescribeRepositories request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

$sel:httpStatus:DescribeRegistriesResponse', describeRegistriesResponse_httpStatus - The response's http status code.

$sel:registries:DescribeRegistriesResponse', describeRegistriesResponse_registries - An object containing the details for a public registry.

Response Lenses

describeRegistriesResponse_nextToken :: Lens' DescribeRegistriesResponse (Maybe Text)

The nextToken value to include in a future DescribeRepositories request. When the results of a DescribeRepositories request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

describeRegistriesResponse_httpStatus :: Lens' DescribeRegistriesResponse Int

The response's http status code.

describeRegistriesResponse_registries :: Lens' DescribeRegistriesResponse [Registry]

An object containing the details for a public registry.