Network.AWS.ECRPublic.PutImage

Description

Creates or updates the image manifest and tags associated with an image.

When an image is pushed and all new image layers have been uploaded, the PutImage API is called once to create or update the image manifest and the tags associated with the image.

This operation is used by the Amazon ECR proxy and is not generally used by customers for pulling and pushing images. In most cases, you should use the docker CLI to pull, tag, and push images.

Creating a Request

data PutImage

See: newPutImage smart constructor.

Constructors

PutImage'

Fields

registryId :: Maybe Text

The AWS account ID associated with the public registry that contains the repository in which to put the image. If you do not specify a registry, the default public registry is assumed.

imageManifestMediaType :: Maybe Text

The media type of the image manifest. If you push an image manifest that does not contain the mediaType field, you must specify the imageManifestMediaType in the request.

imageDigest :: Maybe Text

The image digest of the image manifest corresponding to the image.

imageTag :: Maybe Text

The tag to associate with the image. This parameter is required for images that use the Docker Image Manifest V2 Schema 2 or Open Container Initiative (OCI) formats.

repositoryName :: Text

The name of the repository in which to put the image.

imageManifest :: Text

The image manifest corresponding to the image to be uploaded.

newPutImage

Arguments

:: Text
$sel:repositoryName:PutImage'

-> Text
$sel:imageManifest:PutImage'

-> PutImage

Create a value of PutImage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:registryId:PutImage', putImage_registryId - The AWS account ID associated with the public registry that contains the repository in which to put the image. If you do not specify a registry, the default public registry is assumed.

$sel:imageManifestMediaType:PutImage', putImage_imageManifestMediaType - The media type of the image manifest. If you push an image manifest that does not contain the mediaType field, you must specify the imageManifestMediaType in the request.

$sel:imageDigest:PutImage', putImage_imageDigest - The image digest of the image manifest corresponding to the image.

$sel:imageTag:PutImage', putImage_imageTag - The tag to associate with the image. This parameter is required for images that use the Docker Image Manifest V2 Schema 2 or Open Container Initiative (OCI) formats.

$sel:repositoryName:PutImage', putImage_repositoryName - The name of the repository in which to put the image.

$sel:imageManifest:PutImage', putImage_imageManifest - The image manifest corresponding to the image to be uploaded.

Request Lenses

putImage_registryId :: Lens' PutImage (Maybe Text)

The AWS account ID associated with the public registry that contains the repository in which to put the image. If you do not specify a registry, the default public registry is assumed.

putImage_imageManifestMediaType :: Lens' PutImage (Maybe Text)

The media type of the image manifest. If you push an image manifest that does not contain the mediaType field, you must specify the imageManifestMediaType in the request.

putImage_imageDigest :: Lens' PutImage (Maybe Text)

The image digest of the image manifest corresponding to the image.

putImage_imageTag :: Lens' PutImage (Maybe Text)

The tag to associate with the image. This parameter is required for images that use the Docker Image Manifest V2 Schema 2 or Open Container Initiative (OCI) formats.

putImage_repositoryName :: Lens' PutImage Text

The name of the repository in which to put the image.

putImage_imageManifest :: Lens' PutImage Text

The image manifest corresponding to the image to be uploaded.

Destructuring the Response

data PutImageResponse

See: newPutImageResponse smart constructor.

Constructors

PutImageResponse'

Fields

image :: Maybe Image

Details of the image uploaded.

httpStatus :: Int

The response's http status code.

newPutImageResponse

Arguments

:: Int
$sel:httpStatus:PutImageResponse'

-> PutImageResponse

Create a value of PutImageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:image:PutImageResponse', putImageResponse_image - Details of the image uploaded.

$sel:httpStatus:PutImageResponse', putImageResponse_httpStatus - The response's http status code.

Response Lenses

putImageResponse_image :: Lens' PutImageResponse (Maybe Image)

Details of the image uploaded.

putImageResponse_httpStatus :: Lens' PutImageResponse Int

The response's http status code.