Network.AWS.ECRPublic.Types

Description

Service Configuration

defaultService :: Service

API version 2020-10-30 of the Amazon Elastic Container Registry Public SDK configuration.

Errors

_ImageTagAlreadyExistsException :: AsError a => Getting (First ServiceError) a ServiceError

The specified image is tagged with a tag that already exists. The repository is configured for tag immutability.

_LayersNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified layers could not be found, or the specified layer is not valid for this repository.

_ReferencedImagesNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The manifest list is referencing an image that does not exist.

_InvalidParameterException :: AsError a => Getting (First ServiceError) a ServiceError

The specified parameter is invalid. Review the available parameters for the API request.

_LayerAlreadyExistsException :: AsError a => Getting (First ServiceError) a ServiceError

The image layer already exists in the associated repository.

_ServerException :: AsError a => Getting (First ServiceError) a ServiceError

These errors are usually caused by a server-side issue.

_UnsupportedCommandException :: AsError a => Getting (First ServiceError) a ServiceError

The action is not supported in this Region.

_InvalidLayerException :: AsError a => Getting (First ServiceError) a ServiceError

The layer digest calculation performed by Amazon ECR upon receipt of the image layer does not match the digest specified.

_LayerPartTooSmallException :: AsError a => Getting (First ServiceError) a ServiceError

Layer parts must be at least 5 MiB in size.

_ImageDigestDoesNotMatchException :: AsError a => Getting (First ServiceError) a ServiceError

The specified image digest does not match the digest that Amazon ECR calculated for the image.

_ImageNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The image requested does not exist in the specified repository.

_ImageAlreadyExistsException :: AsError a => Getting (First ServiceError) a ServiceError

The specified image has already been pushed, and there were no changes to the manifest or image tag after the last push.

_RepositoryNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified repository could not be found. Check the spelling of the specified repository and ensure that you are performing operations on the correct registry.

_TooManyTagsException :: AsError a => Getting (First ServiceError) a ServiceError

The list of tags on the repository is over the limit. The maximum number of tags that can be applied to a repository is 50.

_UploadNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The upload could not be found, or the specified upload ID is not valid for this repository.

_InvalidLayerPartException :: AsError a => Getting (First ServiceError) a ServiceError

The layer part size is not valid, or the first byte specified is not consecutive to the last byte of a previous layer part upload.

_InvalidTagParameterException :: AsError a => Getting (First ServiceError) a ServiceError

An invalid parameter has been specified. Tag keys can have a maximum character length of 128 characters, and tag values can have a maximum length of 256 characters.

_RepositoryNotEmptyException :: AsError a => Getting (First ServiceError) a ServiceError

The specified repository contains images. To delete a repository that contains images, you must force the deletion with the force parameter.

_RepositoryAlreadyExistsException :: AsError a => Getting (First ServiceError) a ServiceError

The specified repository already exists in the specified registry.

_RepositoryPolicyNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified repository and registry combination does not have an associated repository policy.

_EmptyUploadException :: AsError a => Getting (First ServiceError) a ServiceError

The specified layer upload does not contain any layer parts.

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The operation did not succeed because it would have exceeded a service limit for your account. For more information, see Amazon ECR Service Quotas in the Amazon Elastic Container Registry User Guide.

_RegistryNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The registry does not exist.

ImageFailureCode

newtype ImageFailureCode

Constructors

ImageFailureCode'

Fields

fromImageFailureCode :: Text

Bundled Patterns

pattern ImageFailureCode_ImageNotFound :: ImageFailureCode

pattern ImageFailureCode_ImageReferencedByManifestList :: ImageFailureCode

pattern ImageFailureCode_ImageTagDoesNotMatchDigest :: ImageFailureCode

pattern ImageFailureCode_InvalidImageDigest :: ImageFailureCode

pattern ImageFailureCode_InvalidImageTag :: ImageFailureCode

pattern ImageFailureCode_KmsError :: ImageFailureCode

pattern ImageFailureCode_MissingDigestAndTag :: ImageFailureCode

LayerAvailability

newtype LayerAvailability

Constructors

LayerAvailability'

Fields

fromLayerAvailability :: Text

Bundled Patterns

pattern LayerAvailability_AVAILABLE :: LayerAvailability

pattern LayerAvailability_UNAVAILABLE :: LayerAvailability

LayerFailureCode

newtype LayerFailureCode

Constructors

LayerFailureCode'

Fields

fromLayerFailureCode :: Text

Bundled Patterns

pattern LayerFailureCode_InvalidLayerDigest :: LayerFailureCode

pattern LayerFailureCode_MissingLayerDigest :: LayerFailureCode

RegistryAliasStatus

newtype RegistryAliasStatus

Constructors

RegistryAliasStatus'

Fields

fromRegistryAliasStatus :: Text

Bundled Patterns

pattern RegistryAliasStatus_ACTIVE :: RegistryAliasStatus

pattern RegistryAliasStatus_PENDING :: RegistryAliasStatus

pattern RegistryAliasStatus_REJECTED :: RegistryAliasStatus

AuthorizationData

data AuthorizationData

An authorization token data object that corresponds to a public registry.

See: newAuthorizationData smart constructor.

Constructors

AuthorizationData'

Fields

expiresAt :: Maybe POSIX

The Unix time in seconds and milliseconds when the authorization token expires. Authorization tokens are valid for 12 hours.

authorizationToken :: Maybe Text

A base64-encoded string that contains authorization data for a public Amazon ECR registry. When the string is decoded, it is presented in the format user:password for public registry authentication using docker login.

newAuthorizationData :: AuthorizationData

Create a value of AuthorizationData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expiresAt:AuthorizationData', authorizationData_expiresAt - The Unix time in seconds and milliseconds when the authorization token expires. Authorization tokens are valid for 12 hours.

$sel:authorizationToken:AuthorizationData', authorizationData_authorizationToken - A base64-encoded string that contains authorization data for a public Amazon ECR registry. When the string is decoded, it is presented in the format user:password for public registry authentication using docker login.

authorizationData_expiresAt :: Lens' AuthorizationData (Maybe UTCTime)

The Unix time in seconds and milliseconds when the authorization token expires. Authorization tokens are valid for 12 hours.

authorizationData_authorizationToken :: Lens' AuthorizationData (Maybe Text)

A base64-encoded string that contains authorization data for a public Amazon ECR registry. When the string is decoded, it is presented in the format user:password for public registry authentication using docker login.

Image

data Image

An object representing an Amazon ECR image.

See: newImage smart constructor.

Constructors

Image'

Fields

registryId :: Maybe Text

The AWS account ID associated with the registry containing the image.

imageManifestMediaType :: Maybe Text

The manifest media type of the image.

imageId :: Maybe ImageIdentifier

An object containing the image tag and image digest associated with an image.

repositoryName :: Maybe Text

The name of the repository associated with the image.

imageManifest :: Maybe Text

The image manifest associated with the image.

newImage :: Image

Create a value of Image with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:registryId:Image', image_registryId - The AWS account ID associated with the registry containing the image.

$sel:imageManifestMediaType:Image', image_imageManifestMediaType - The manifest media type of the image.

$sel:imageId:Image', image_imageId - An object containing the image tag and image digest associated with an image.

$sel:repositoryName:Image', image_repositoryName - The name of the repository associated with the image.

$sel:imageManifest:Image', image_imageManifest - The image manifest associated with the image.

image_registryId :: Lens' Image (Maybe Text)

The AWS account ID associated with the registry containing the image.

image_imageManifestMediaType :: Lens' Image (Maybe Text)

The manifest media type of the image.

image_imageId :: Lens' Image (Maybe ImageIdentifier)

An object containing the image tag and image digest associated with an image.

image_repositoryName :: Lens' Image (Maybe Text)

The name of the repository associated with the image.

image_imageManifest :: Lens' Image (Maybe Text)

The image manifest associated with the image.

ImageDetail

data ImageDetail

An object that describes an image returned by a DescribeImages operation.

See: newImageDetail smart constructor.

Constructors

ImageDetail'

Fields

registryId :: Maybe Text

The AWS account ID associated with the public registry to which this image belongs.

imageTags :: Maybe [Text]

The list of tags associated with this image.

imageManifestMediaType :: Maybe Text

The media type of the image manifest.

imageSizeInBytes :: Maybe Integer

The size, in bytes, of the image in the repository.

If the image is a manifest list, this will be the max size of all manifests in the list.

Beginning with Docker version 1.9, the Docker client compresses image layers before pushing them to a V2 Docker registry. The output of the docker images command shows the uncompressed image size, so it may return a larger image size than the image sizes returned by DescribeImages.

imageDigest :: Maybe Text

The sha256 digest of the image manifest.

artifactMediaType :: Maybe Text

The artifact media type of the image.

imagePushedAt :: Maybe POSIX

The date and time, expressed in standard JavaScript date format, at which the current image was pushed to the repository.

repositoryName :: Maybe Text

The name of the repository to which this image belongs.

newImageDetail :: ImageDetail

Create a value of ImageDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:registryId:ImageDetail', imageDetail_registryId - The AWS account ID associated with the public registry to which this image belongs.

$sel:imageTags:ImageDetail', imageDetail_imageTags - The list of tags associated with this image.

$sel:imageManifestMediaType:ImageDetail', imageDetail_imageManifestMediaType - The media type of the image manifest.

$sel:imageSizeInBytes:ImageDetail', imageDetail_imageSizeInBytes - The size, in bytes, of the image in the repository.

If the image is a manifest list, this will be the max size of all manifests in the list.

Beginning with Docker version 1.9, the Docker client compresses image layers before pushing them to a V2 Docker registry. The output of the docker images command shows the uncompressed image size, so it may return a larger image size than the image sizes returned by DescribeImages.

$sel:imageDigest:ImageDetail', imageDetail_imageDigest - The sha256 digest of the image manifest.

$sel:artifactMediaType:ImageDetail', imageDetail_artifactMediaType - The artifact media type of the image.

$sel:imagePushedAt:ImageDetail', imageDetail_imagePushedAt - The date and time, expressed in standard JavaScript date format, at which the current image was pushed to the repository.

$sel:repositoryName:ImageDetail', imageDetail_repositoryName - The name of the repository to which this image belongs.

imageDetail_registryId :: Lens' ImageDetail (Maybe Text)

The AWS account ID associated with the public registry to which this image belongs.

imageDetail_imageTags :: Lens' ImageDetail (Maybe [Text])

The list of tags associated with this image.

imageDetail_imageManifestMediaType :: Lens' ImageDetail (Maybe Text)

The media type of the image manifest.

imageDetail_imageSizeInBytes :: Lens' ImageDetail (Maybe Integer)

The size, in bytes, of the image in the repository.

If the image is a manifest list, this will be the max size of all manifests in the list.

Beginning with Docker version 1.9, the Docker client compresses image layers before pushing them to a V2 Docker registry. The output of the docker images command shows the uncompressed image size, so it may return a larger image size than the image sizes returned by DescribeImages.

imageDetail_imageDigest :: Lens' ImageDetail (Maybe Text)

The sha256 digest of the image manifest.

imageDetail_artifactMediaType :: Lens' ImageDetail (Maybe Text)

The artifact media type of the image.

imageDetail_imagePushedAt :: Lens' ImageDetail (Maybe UTCTime)

The date and time, expressed in standard JavaScript date format, at which the current image was pushed to the repository.

imageDetail_repositoryName :: Lens' ImageDetail (Maybe Text)

The name of the repository to which this image belongs.

ImageFailure

data ImageFailure

An object representing an Amazon ECR image failure.

See: newImageFailure smart constructor.

Constructors

ImageFailure'

Fields

failureReason :: Maybe Text

The reason for the failure.

failureCode :: Maybe ImageFailureCode

The code associated with the failure.

imageId :: Maybe ImageIdentifier

The image ID associated with the failure.

newImageFailure :: ImageFailure

Create a value of ImageFailure with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:ImageFailure', imageFailure_failureReason - The reason for the failure.

$sel:failureCode:ImageFailure', imageFailure_failureCode - The code associated with the failure.

$sel:imageId:ImageFailure', imageFailure_imageId - The image ID associated with the failure.

imageFailure_failureReason :: Lens' ImageFailure (Maybe Text)

The reason for the failure.

imageFailure_failureCode :: Lens' ImageFailure (Maybe ImageFailureCode)

The code associated with the failure.

imageFailure_imageId :: Lens' ImageFailure (Maybe ImageIdentifier)

The image ID associated with the failure.

ImageIdentifier

data ImageIdentifier

An object with identifying information for an Amazon ECR image.

See: newImageIdentifier smart constructor.

Constructors

ImageIdentifier'

Fields

imageDigest :: Maybe Text

The sha256 digest of the image manifest.

imageTag :: Maybe Text

The tag used for the image.

newImageIdentifier :: ImageIdentifier

Create a value of ImageIdentifier with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:imageDigest:ImageIdentifier', imageIdentifier_imageDigest - The sha256 digest of the image manifest.

$sel:imageTag:ImageIdentifier', imageIdentifier_imageTag - The tag used for the image.

imageIdentifier_imageDigest :: Lens' ImageIdentifier (Maybe Text)

The sha256 digest of the image manifest.

imageIdentifier_imageTag :: Lens' ImageIdentifier (Maybe Text)

The tag used for the image.

ImageTagDetail

data ImageTagDetail

An object representing the image tag details for an image.

See: newImageTagDetail smart constructor.

Constructors

ImageTagDetail'

Fields

createdAt :: Maybe POSIX

The time stamp indicating when the image tag was created.

imageDetail :: Maybe ReferencedImageDetail

An object that describes the details of an image.

imageTag :: Maybe Text

The tag associated with the image.

newImageTagDetail :: ImageTagDetail

Create a value of ImageTagDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createdAt:ImageTagDetail', imageTagDetail_createdAt - The time stamp indicating when the image tag was created.

$sel:imageDetail:ImageTagDetail', imageTagDetail_imageDetail - An object that describes the details of an image.

$sel:imageTag:ImageTagDetail', imageTagDetail_imageTag - The tag associated with the image.

imageTagDetail_createdAt :: Lens' ImageTagDetail (Maybe UTCTime)

The time stamp indicating when the image tag was created.

imageTagDetail_imageDetail :: Lens' ImageTagDetail (Maybe ReferencedImageDetail)

An object that describes the details of an image.

imageTagDetail_imageTag :: Lens' ImageTagDetail (Maybe Text)

The tag associated with the image.

Layer

data Layer

An object representing an Amazon ECR image layer.

See: newLayer smart constructor.

Constructors

Layer'

Fields

mediaType :: Maybe Text

The media type of the layer, such as application/vnd.docker.image.rootfs.diff.tar.gzip or application/vnd.oci.image.layer.v1.tar+gzip.

layerDigest :: Maybe Text

The sha256 digest of the image layer.

layerSize :: Maybe Integer

The size, in bytes, of the image layer.

layerAvailability :: Maybe LayerAvailability

The availability status of the image layer.

newLayer :: Layer

Create a value of Layer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:mediaType:Layer', layer_mediaType - The media type of the layer, such as application/vnd.docker.image.rootfs.diff.tar.gzip or application/vnd.oci.image.layer.v1.tar+gzip.

$sel:layerDigest:Layer', layer_layerDigest - The sha256 digest of the image layer.

$sel:layerSize:Layer', layer_layerSize - The size, in bytes, of the image layer.

$sel:layerAvailability:Layer', layer_layerAvailability - The availability status of the image layer.

layer_mediaType :: Lens' Layer (Maybe Text)

The media type of the layer, such as application/vnd.docker.image.rootfs.diff.tar.gzip or application/vnd.oci.image.layer.v1.tar+gzip.

layer_layerDigest :: Lens' Layer (Maybe Text)

The sha256 digest of the image layer.

layer_layerSize :: Lens' Layer (Maybe Integer)

The size, in bytes, of the image layer.

layer_layerAvailability :: Lens' Layer (Maybe LayerAvailability)

The availability status of the image layer.

LayerFailure

data LayerFailure

An object representing an Amazon ECR image layer failure.

See: newLayerFailure smart constructor.

Constructors

LayerFailure'

Fields

failureReason :: Maybe Text

The reason for the failure.

failureCode :: Maybe LayerFailureCode

The failure code associated with the failure.

layerDigest :: Maybe Text

The layer digest associated with the failure.

newLayerFailure :: LayerFailure

Create a value of LayerFailure with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:LayerFailure', layerFailure_failureReason - The reason for the failure.

$sel:failureCode:LayerFailure', layerFailure_failureCode - The failure code associated with the failure.

$sel:layerDigest:LayerFailure', layerFailure_layerDigest - The layer digest associated with the failure.

layerFailure_failureReason :: Lens' LayerFailure (Maybe Text)

The reason for the failure.

layerFailure_failureCode :: Lens' LayerFailure (Maybe LayerFailureCode)

The failure code associated with the failure.

layerFailure_layerDigest :: Lens' LayerFailure (Maybe Text)

The layer digest associated with the failure.

ReferencedImageDetail

data ReferencedImageDetail

An object that describes the image tag details returned by a DescribeImageTags action.

See: newReferencedImageDetail smart constructor.

Constructors

ReferencedImageDetail'

Fields

imageManifestMediaType :: Maybe Text

The media type of the image manifest.

imageSizeInBytes :: Maybe Integer

The size, in bytes, of the image in the repository.

If the image is a manifest list, this will be the max size of all manifests in the list.

Beginning with Docker version 1.9, the Docker client compresses image layers before pushing them to a V2 Docker registry. The output of the docker images command shows the uncompressed image size, so it may return a larger image size than the image sizes returned by DescribeImages.

imageDigest :: Maybe Text

The sha256 digest of the image manifest.

artifactMediaType :: Maybe Text

The artifact media type of the image.

imagePushedAt :: Maybe POSIX

The date and time, expressed in standard JavaScript date format, at which the current image tag was pushed to the repository.

newReferencedImageDetail :: ReferencedImageDetail

Create a value of ReferencedImageDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:imageManifestMediaType:ReferencedImageDetail', referencedImageDetail_imageManifestMediaType - The media type of the image manifest.

$sel:imageSizeInBytes:ReferencedImageDetail', referencedImageDetail_imageSizeInBytes - The size, in bytes, of the image in the repository.

If the image is a manifest list, this will be the max size of all manifests in the list.

Beginning with Docker version 1.9, the Docker client compresses image layers before pushing them to a V2 Docker registry. The output of the docker images command shows the uncompressed image size, so it may return a larger image size than the image sizes returned by DescribeImages.

$sel:imageDigest:ReferencedImageDetail', referencedImageDetail_imageDigest - The sha256 digest of the image manifest.

$sel:artifactMediaType:ReferencedImageDetail', referencedImageDetail_artifactMediaType - The artifact media type of the image.

$sel:imagePushedAt:ReferencedImageDetail', referencedImageDetail_imagePushedAt - The date and time, expressed in standard JavaScript date format, at which the current image tag was pushed to the repository.

referencedImageDetail_imageManifestMediaType :: Lens' ReferencedImageDetail (Maybe Text)

The media type of the image manifest.

referencedImageDetail_imageSizeInBytes :: Lens' ReferencedImageDetail (Maybe Integer)

The size, in bytes, of the image in the repository.

If the image is a manifest list, this will be the max size of all manifests in the list.

Beginning with Docker version 1.9, the Docker client compresses image layers before pushing them to a V2 Docker registry. The output of the docker images command shows the uncompressed image size, so it may return a larger image size than the image sizes returned by DescribeImages.

referencedImageDetail_imageDigest :: Lens' ReferencedImageDetail (Maybe Text)

The sha256 digest of the image manifest.

referencedImageDetail_artifactMediaType :: Lens' ReferencedImageDetail (Maybe Text)

The artifact media type of the image.

referencedImageDetail_imagePushedAt :: Lens' ReferencedImageDetail (Maybe UTCTime)

The date and time, expressed in standard JavaScript date format, at which the current image tag was pushed to the repository.

Registry

data Registry

The details of a public registry.

See: newRegistry smart constructor.

Constructors

Registry'

Fields

registryId :: Text

The AWS account ID associated with the registry. If you do not specify a registry, the default public registry is assumed.

registryArn :: Text

The Amazon Resource Name (ARN) of the public registry.

registryUri :: Text

The URI of a public registry. The URI contains a universal prefix and the registry alias.

verified :: Bool

Whether the account is verified. This indicates whether the account is an AWS Marketplace vendor. If an account is verified, each public repository will received a verified account badge on the Amazon ECR Public Gallery.

aliases :: [RegistryAlias]

An array of objects representing the aliases for a public registry.

newRegistry

Arguments

:: Text
$sel:registryId:Registry'

-> Text
$sel:registryArn:Registry'

-> Text
$sel:registryUri:Registry'

-> Bool
$sel:verified:Registry'

-> Registry

Create a value of Registry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:registryId:Registry', registry_registryId - The AWS account ID associated with the registry. If you do not specify a registry, the default public registry is assumed.

$sel:registryArn:Registry', registry_registryArn - The Amazon Resource Name (ARN) of the public registry.

$sel:registryUri:Registry', registry_registryUri - The URI of a public registry. The URI contains a universal prefix and the registry alias.

$sel:verified:Registry', registry_verified - Whether the account is verified. This indicates whether the account is an AWS Marketplace vendor. If an account is verified, each public repository will received a verified account badge on the Amazon ECR Public Gallery.

$sel:aliases:Registry', registry_aliases - An array of objects representing the aliases for a public registry.

registry_registryId :: Lens' Registry Text

The AWS account ID associated with the registry. If you do not specify a registry, the default public registry is assumed.

registry_registryArn :: Lens' Registry Text

The Amazon Resource Name (ARN) of the public registry.

registry_registryUri :: Lens' Registry Text

The URI of a public registry. The URI contains a universal prefix and the registry alias.

registry_verified :: Lens' Registry Bool

Whether the account is verified. This indicates whether the account is an AWS Marketplace vendor. If an account is verified, each public repository will received a verified account badge on the Amazon ECR Public Gallery.

registry_aliases :: Lens' Registry [RegistryAlias]

An array of objects representing the aliases for a public registry.

RegistryAlias

data RegistryAlias

An object representing the aliases for a public registry. A public registry is given an alias upon creation but a custom alias can be set using the Amazon ECR console. For more information, see Registries in the Amazon Elastic Container Registry User Guide.

See: newRegistryAlias smart constructor.

Constructors

RegistryAlias'

Fields

name :: Text

The name of the registry alias.

status :: RegistryAliasStatus

The status of the registry alias.

primaryRegistryAlias :: Bool

Whether or not the registry alias is the primary alias for the registry. If true, the alias is the primary registry alias and is displayed in both the repository URL and the image URI used in the docker pull commands on the Amazon ECR Public Gallery.

A registry alias that is not the primary registry alias can be used in the repository URI in a docker pull command.

defaultRegistryAlias :: Bool

Whether or not the registry alias is the default alias for the registry. When the first public repository is created, your public registry is assigned a default registry alias.

newRegistryAlias

Arguments

:: Text
$sel:name:RegistryAlias'

-> RegistryAliasStatus
$sel:status:RegistryAlias'

-> Bool
$sel:primaryRegistryAlias:RegistryAlias'

-> Bool
$sel:defaultRegistryAlias:RegistryAlias'

-> RegistryAlias

Create a value of RegistryAlias with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:RegistryAlias', registryAlias_name - The name of the registry alias.

$sel:status:RegistryAlias', registryAlias_status - The status of the registry alias.

$sel:primaryRegistryAlias:RegistryAlias', registryAlias_primaryRegistryAlias - Whether or not the registry alias is the primary alias for the registry. If true, the alias is the primary registry alias and is displayed in both the repository URL and the image URI used in the docker pull commands on the Amazon ECR Public Gallery.

A registry alias that is not the primary registry alias can be used in the repository URI in a docker pull command.

$sel:defaultRegistryAlias:RegistryAlias', registryAlias_defaultRegistryAlias - Whether or not the registry alias is the default alias for the registry. When the first public repository is created, your public registry is assigned a default registry alias.

registryAlias_name :: Lens' RegistryAlias Text

The name of the registry alias.

registryAlias_status :: Lens' RegistryAlias RegistryAliasStatus

The status of the registry alias.

registryAlias_primaryRegistryAlias :: Lens' RegistryAlias Bool

Whether or not the registry alias is the primary alias for the registry. If true, the alias is the primary registry alias and is displayed in both the repository URL and the image URI used in the docker pull commands on the Amazon ECR Public Gallery.

A registry alias that is not the primary registry alias can be used in the repository URI in a docker pull command.

registryAlias_defaultRegistryAlias :: Lens' RegistryAlias Bool

Whether or not the registry alias is the default alias for the registry. When the first public repository is created, your public registry is assigned a default registry alias.

RegistryCatalogData

data RegistryCatalogData

The metadata for a public registry.

See: newRegistryCatalogData smart constructor.

Constructors

RegistryCatalogData'

Fields

displayName :: Maybe Text

The display name for a public registry. This appears on the Amazon ECR Public Gallery.

Only accounts that have the verified account badge can have a registry display name.

newRegistryCatalogData :: RegistryCatalogData

Create a value of RegistryCatalogData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:displayName:RegistryCatalogData', registryCatalogData_displayName - The display name for a public registry. This appears on the Amazon ECR Public Gallery.

Only accounts that have the verified account badge can have a registry display name.

registryCatalogData_displayName :: Lens' RegistryCatalogData (Maybe Text)

The display name for a public registry. This appears on the Amazon ECR Public Gallery.

Only accounts that have the verified account badge can have a registry display name.

Repository

data Repository

An object representing a repository.

See: newRepository smart constructor.

Constructors

Repository'

Fields

repositoryArn :: Maybe Text

The Amazon Resource Name (ARN) that identifies the repository. The ARN contains the arn:aws:ecr namespace, followed by the region of the repository, AWS account ID of the repository owner, repository namespace, and repository name. For example, arn:aws:ecr:region:012345678910:repository/test.

createdAt :: Maybe POSIX

The date and time, in JavaScript date format, when the repository was created.

registryId :: Maybe Text

The AWS account ID associated with the public registry that contains the repository.

repositoryUri :: Maybe Text

The URI for the repository. You can use this URI for container image push and pull operations.

repositoryName :: Maybe Text

The name of the repository.

newRepository :: Repository

Create a value of Repository with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:repositoryArn:Repository', repository_repositoryArn - The Amazon Resource Name (ARN) that identifies the repository. The ARN contains the arn:aws:ecr namespace, followed by the region of the repository, AWS account ID of the repository owner, repository namespace, and repository name. For example, arn:aws:ecr:region:012345678910:repository/test.

$sel:createdAt:Repository', repository_createdAt - The date and time, in JavaScript date format, when the repository was created.

$sel:registryId:Repository', repository_registryId - The AWS account ID associated with the public registry that contains the repository.

$sel:repositoryUri:Repository', repository_repositoryUri - The URI for the repository. You can use this URI for container image push and pull operations.

$sel:repositoryName:Repository', repository_repositoryName - The name of the repository.

repository_repositoryArn :: Lens' Repository (Maybe Text)

The Amazon Resource Name (ARN) that identifies the repository. The ARN contains the arn:aws:ecr namespace, followed by the region of the repository, AWS account ID of the repository owner, repository namespace, and repository name. For example, arn:aws:ecr:region:012345678910:repository/test.

repository_createdAt :: Lens' Repository (Maybe UTCTime)

The date and time, in JavaScript date format, when the repository was created.

repository_registryId :: Lens' Repository (Maybe Text)

The AWS account ID associated with the public registry that contains the repository.

repository_repositoryUri :: Lens' Repository (Maybe Text)

The URI for the repository. You can use this URI for container image push and pull operations.

repository_repositoryName :: Lens' Repository (Maybe Text)

The name of the repository.

RepositoryCatalogData

data RepositoryCatalogData

The catalog data for a repository. This data is publicly visible in the Amazon ECR Public Gallery.

See: newRepositoryCatalogData smart constructor.

Constructors

RepositoryCatalogData'

Fields

logoUrl :: Maybe Text

The URL containing the logo associated with the repository.

architectures :: Maybe [Text]

The architecture tags that are associated with the repository.

Only supported operating system tags appear publicly in the Amazon ECR Public Gallery. For more information, see RepositoryCatalogDataInput.

usageText :: Maybe Text

The longform usage details of the contents of the repository. The usage text provides context for users of the repository.

marketplaceCertified :: Maybe Bool

Whether or not the repository is certified by AWS Marketplace.

aboutText :: Maybe Text

The longform description of the contents of the repository. This text appears in the repository details on the Amazon ECR Public Gallery.

operatingSystems :: Maybe [Text]

The operating system tags that are associated with the repository.

Only supported operating system tags appear publicly in the Amazon ECR Public Gallery. For more information, see RepositoryCatalogDataInput.

description :: Maybe Text

The short description of the repository.

newRepositoryCatalogData :: RepositoryCatalogData

Create a value of RepositoryCatalogData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:logoUrl:RepositoryCatalogData', repositoryCatalogData_logoUrl - The URL containing the logo associated with the repository.

$sel:architectures:RepositoryCatalogData', repositoryCatalogData_architectures - The architecture tags that are associated with the repository.

Only supported operating system tags appear publicly in the Amazon ECR Public Gallery. For more information, see RepositoryCatalogDataInput.

$sel:usageText:RepositoryCatalogData', repositoryCatalogData_usageText - The longform usage details of the contents of the repository. The usage text provides context for users of the repository.

$sel:marketplaceCertified:RepositoryCatalogData', repositoryCatalogData_marketplaceCertified - Whether or not the repository is certified by AWS Marketplace.

$sel:aboutText:RepositoryCatalogData', repositoryCatalogData_aboutText - The longform description of the contents of the repository. This text appears in the repository details on the Amazon ECR Public Gallery.

$sel:operatingSystems:RepositoryCatalogData', repositoryCatalogData_operatingSystems - The operating system tags that are associated with the repository.

Only supported operating system tags appear publicly in the Amazon ECR Public Gallery. For more information, see RepositoryCatalogDataInput.

$sel:description:RepositoryCatalogData', repositoryCatalogData_description - The short description of the repository.

repositoryCatalogData_logoUrl :: Lens' RepositoryCatalogData (Maybe Text)

The URL containing the logo associated with the repository.

repositoryCatalogData_architectures :: Lens' RepositoryCatalogData (Maybe [Text])

The architecture tags that are associated with the repository.

Only supported operating system tags appear publicly in the Amazon ECR Public Gallery. For more information, see RepositoryCatalogDataInput.

repositoryCatalogData_usageText :: Lens' RepositoryCatalogData (Maybe Text)

The longform usage details of the contents of the repository. The usage text provides context for users of the repository.

repositoryCatalogData_marketplaceCertified :: Lens' RepositoryCatalogData (Maybe Bool)

Whether or not the repository is certified by AWS Marketplace.

repositoryCatalogData_aboutText :: Lens' RepositoryCatalogData (Maybe Text)

The longform description of the contents of the repository. This text appears in the repository details on the Amazon ECR Public Gallery.

repositoryCatalogData_operatingSystems :: Lens' RepositoryCatalogData (Maybe [Text])

The operating system tags that are associated with the repository.

Only supported operating system tags appear publicly in the Amazon ECR Public Gallery. For more information, see RepositoryCatalogDataInput.

repositoryCatalogData_description :: Lens' RepositoryCatalogData (Maybe Text)

The short description of the repository.

RepositoryCatalogDataInput

data RepositoryCatalogDataInput

An object containing the catalog data for a repository. This data is publicly visible in the Amazon ECR Public Gallery.

See: newRepositoryCatalogDataInput smart constructor.

Constructors

RepositoryCatalogDataInput'

Fields

logoImageBlob :: Maybe Base64

The base64-encoded repository logo payload.

The repository logo is only publicly visible in the Amazon ECR Public Gallery for verified accounts.

architectures :: Maybe [Text]

The system architecture that the images in the repository are compatible with. On the Amazon ECR Public Gallery, the following supported architectures will appear as badges on the repository and are used as search filters.

Linux

Windows

If an unsupported tag is added to your repository catalog data, it will be associated with the repository and can be retrieved using the API but will not be discoverable in the Amazon ECR Public Gallery.

usageText :: Maybe Text

Detailed information on how to use the contents of the repository. It is publicly visible in the Amazon ECR Public Gallery. The usage text provides context, support information, and additional usage details for users of the repository. The text must be in markdown format.

aboutText :: Maybe Text

A detailed description of the contents of the repository. It is publicly visible in the Amazon ECR Public Gallery. The text must be in markdown format.

operatingSystems :: Maybe [Text]

The operating systems that the images in the repository are compatible with. On the Amazon ECR Public Gallery, the following supported operating systems will appear as badges on the repository and are used as search filters.

ARM

ARM 64

x86

x86-64

If an unsupported tag is added to your repository catalog data, it will be associated with the repository and can be retrieved using the API but will not be discoverable in the Amazon ECR Public Gallery.

description :: Maybe Text

A short description of the contents of the repository. This text appears in both the image details and also when searching for repositories on the Amazon ECR Public Gallery.

newRepositoryCatalogDataInput :: RepositoryCatalogDataInput

Create a value of RepositoryCatalogDataInput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:logoImageBlob:RepositoryCatalogDataInput', repositoryCatalogDataInput_logoImageBlob - The base64-encoded repository logo payload.

The repository logo is only publicly visible in the Amazon ECR Public Gallery for verified accounts.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:architectures:RepositoryCatalogDataInput', repositoryCatalogDataInput_architectures - The system architecture that the images in the repository are compatible with. On the Amazon ECR Public Gallery, the following supported architectures will appear as badges on the repository and are used as search filters.

Linux

Windows

If an unsupported tag is added to your repository catalog data, it will be associated with the repository and can be retrieved using the API but will not be discoverable in the Amazon ECR Public Gallery.

$sel:usageText:RepositoryCatalogDataInput', repositoryCatalogDataInput_usageText - Detailed information on how to use the contents of the repository. It is publicly visible in the Amazon ECR Public Gallery. The usage text provides context, support information, and additional usage details for users of the repository. The text must be in markdown format.

$sel:aboutText:RepositoryCatalogDataInput', repositoryCatalogDataInput_aboutText - A detailed description of the contents of the repository. It is publicly visible in the Amazon ECR Public Gallery. The text must be in markdown format.

$sel:operatingSystems:RepositoryCatalogDataInput', repositoryCatalogDataInput_operatingSystems - The operating systems that the images in the repository are compatible with. On the Amazon ECR Public Gallery, the following supported operating systems will appear as badges on the repository and are used as search filters.

ARM

ARM 64

x86

x86-64

If an unsupported tag is added to your repository catalog data, it will be associated with the repository and can be retrieved using the API but will not be discoverable in the Amazon ECR Public Gallery.

$sel:description:RepositoryCatalogDataInput', repositoryCatalogDataInput_description - A short description of the contents of the repository. This text appears in both the image details and also when searching for repositories on the Amazon ECR Public Gallery.

repositoryCatalogDataInput_logoImageBlob :: Lens' RepositoryCatalogDataInput (Maybe ByteString)

The base64-encoded repository logo payload.

The repository logo is only publicly visible in the Amazon ECR Public Gallery for verified accounts.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

repositoryCatalogDataInput_architectures :: Lens' RepositoryCatalogDataInput (Maybe [Text])

The system architecture that the images in the repository are compatible with. On the Amazon ECR Public Gallery, the following supported architectures will appear as badges on the repository and are used as search filters.

Linux

Windows

If an unsupported tag is added to your repository catalog data, it will be associated with the repository and can be retrieved using the API but will not be discoverable in the Amazon ECR Public Gallery.

repositoryCatalogDataInput_usageText :: Lens' RepositoryCatalogDataInput (Maybe Text)

Detailed information on how to use the contents of the repository. It is publicly visible in the Amazon ECR Public Gallery. The usage text provides context, support information, and additional usage details for users of the repository. The text must be in markdown format.

repositoryCatalogDataInput_aboutText :: Lens' RepositoryCatalogDataInput (Maybe Text)

A detailed description of the contents of the repository. It is publicly visible in the Amazon ECR Public Gallery. The text must be in markdown format.

repositoryCatalogDataInput_operatingSystems :: Lens' RepositoryCatalogDataInput (Maybe [Text])

The operating systems that the images in the repository are compatible with. On the Amazon ECR Public Gallery, the following supported operating systems will appear as badges on the repository and are used as search filters.

ARM

ARM 64

x86

x86-64

If an unsupported tag is added to your repository catalog data, it will be associated with the repository and can be retrieved using the API but will not be discoverable in the Amazon ECR Public Gallery.

repositoryCatalogDataInput_description :: Lens' RepositoryCatalogDataInput (Maybe Text)

A short description of the contents of the repository. This text appears in both the image details and also when searching for repositories on the Amazon ECR Public Gallery.

Tag

data Tag

The metadata that you apply to a resource to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define. Tag keys can have a maximum character length of 128 characters, and tag values can have a maximum length of 256 characters.

See: newTag smart constructor.

Constructors

Tag'

Fields

value :: Maybe Text

The optional part of a key-value pair that make up a tag. A value acts as a descriptor within a tag category (key).

key :: Maybe Text

One part of a key-value pair that make up a tag. A key is a general label that acts like a category for more specific tag values.

newTag :: Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:Tag', tag_value - The optional part of a key-value pair that make up a tag. A value acts as a descriptor within a tag category (key).

$sel:key:Tag', tag_key - One part of a key-value pair that make up a tag. A key is a general label that acts like a category for more specific tag values.

tag_value :: Lens' Tag (Maybe Text)

The optional part of a key-value pair that make up a tag. A value acts as a descriptor within a tag category (key).

tag_key :: Lens' Tag (Maybe Text)

One part of a key-value pair that make up a tag. A key is a general label that acts like a category for more specific tag values.