Network.AWS.ECRPublic.Types.ReferencedImageDetail

Description

Documentation

data ReferencedImageDetail

An object that describes the image tag details returned by a DescribeImageTags action.

See: newReferencedImageDetail smart constructor.

Constructors

ReferencedImageDetail'

Fields

imageManifestMediaType :: Maybe Text

The media type of the image manifest.

imageSizeInBytes :: Maybe Integer

The size, in bytes, of the image in the repository.

If the image is a manifest list, this will be the max size of all manifests in the list.

Beginning with Docker version 1.9, the Docker client compresses image layers before pushing them to a V2 Docker registry. The output of the docker images command shows the uncompressed image size, so it may return a larger image size than the image sizes returned by DescribeImages.

imageDigest :: Maybe Text

The sha256 digest of the image manifest.

artifactMediaType :: Maybe Text

The artifact media type of the image.

imagePushedAt :: Maybe POSIX

The date and time, expressed in standard JavaScript date format, at which the current image tag was pushed to the repository.

newReferencedImageDetail :: ReferencedImageDetail

Create a value of ReferencedImageDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:imageManifestMediaType:ReferencedImageDetail', referencedImageDetail_imageManifestMediaType - The media type of the image manifest.

$sel:imageSizeInBytes:ReferencedImageDetail', referencedImageDetail_imageSizeInBytes - The size, in bytes, of the image in the repository.

If the image is a manifest list, this will be the max size of all manifests in the list.

Beginning with Docker version 1.9, the Docker client compresses image layers before pushing them to a V2 Docker registry. The output of the docker images command shows the uncompressed image size, so it may return a larger image size than the image sizes returned by DescribeImages.

$sel:imageDigest:ReferencedImageDetail', referencedImageDetail_imageDigest - The sha256 digest of the image manifest.

$sel:artifactMediaType:ReferencedImageDetail', referencedImageDetail_artifactMediaType - The artifact media type of the image.

$sel:imagePushedAt:ReferencedImageDetail', referencedImageDetail_imagePushedAt - The date and time, expressed in standard JavaScript date format, at which the current image tag was pushed to the repository.

referencedImageDetail_imageManifestMediaType :: Lens' ReferencedImageDetail (Maybe Text)

The media type of the image manifest.

referencedImageDetail_imageSizeInBytes :: Lens' ReferencedImageDetail (Maybe Integer)

The size, in bytes, of the image in the repository.

If the image is a manifest list, this will be the max size of all manifests in the list.

Beginning with Docker version 1.9, the Docker client compresses image layers before pushing them to a V2 Docker registry. The output of the docker images command shows the uncompressed image size, so it may return a larger image size than the image sizes returned by DescribeImages.

referencedImageDetail_imageDigest :: Lens' ReferencedImageDetail (Maybe Text)

The sha256 digest of the image manifest.

referencedImageDetail_artifactMediaType :: Lens' ReferencedImageDetail (Maybe Text)

The artifact media type of the image.

referencedImageDetail_imagePushedAt :: Lens' ReferencedImageDetail (Maybe UTCTime)

The date and time, expressed in standard JavaScript date format, at which the current image tag was pushed to the repository.