Network.AWS.ECRPublic.Types.RegistryAlias

Description

Documentation

data RegistryAlias

An object representing the aliases for a public registry. A public registry is given an alias upon creation but a custom alias can be set using the Amazon ECR console. For more information, see Registries in the Amazon Elastic Container Registry User Guide.

See: newRegistryAlias smart constructor.

Constructors

RegistryAlias'

Fields

name :: Text

The name of the registry alias.

status :: RegistryAliasStatus

The status of the registry alias.

primaryRegistryAlias :: Bool

Whether or not the registry alias is the primary alias for the registry. If true, the alias is the primary registry alias and is displayed in both the repository URL and the image URI used in the docker pull commands on the Amazon ECR Public Gallery.

A registry alias that is not the primary registry alias can be used in the repository URI in a docker pull command.

defaultRegistryAlias :: Bool

Whether or not the registry alias is the default alias for the registry. When the first public repository is created, your public registry is assigned a default registry alias.

newRegistryAlias

Arguments

:: Text
$sel:name:RegistryAlias'

-> RegistryAliasStatus
$sel:status:RegistryAlias'

-> Bool
$sel:primaryRegistryAlias:RegistryAlias'

-> Bool
$sel:defaultRegistryAlias:RegistryAlias'

-> RegistryAlias

Create a value of RegistryAlias with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:RegistryAlias', registryAlias_name - The name of the registry alias.

$sel:status:RegistryAlias', registryAlias_status - The status of the registry alias.

$sel:primaryRegistryAlias:RegistryAlias', registryAlias_primaryRegistryAlias - Whether or not the registry alias is the primary alias for the registry. If true, the alias is the primary registry alias and is displayed in both the repository URL and the image URI used in the docker pull commands on the Amazon ECR Public Gallery.

A registry alias that is not the primary registry alias can be used in the repository URI in a docker pull command.

$sel:defaultRegistryAlias:RegistryAlias', registryAlias_defaultRegistryAlias - Whether or not the registry alias is the default alias for the registry. When the first public repository is created, your public registry is assigned a default registry alias.

registryAlias_name :: Lens' RegistryAlias Text

The name of the registry alias.

registryAlias_status :: Lens' RegistryAlias RegistryAliasStatus

The status of the registry alias.

registryAlias_primaryRegistryAlias :: Lens' RegistryAlias Bool

Whether or not the registry alias is the primary alias for the registry. If true, the alias is the primary registry alias and is displayed in both the repository URL and the image URI used in the docker pull commands on the Amazon ECR Public Gallery.

A registry alias that is not the primary registry alias can be used in the repository URI in a docker pull command.

registryAlias_defaultRegistryAlias :: Lens' RegistryAlias Bool

Whether or not the registry alias is the default alias for the registry. When the first public repository is created, your public registry is assigned a default registry alias.