Network.AWS.ECRPublic.Types.RegistryCatalogData

Description

Documentation

data RegistryCatalogData

The metadata for a public registry.

See: newRegistryCatalogData smart constructor.

Constructors

RegistryCatalogData'

Fields

displayName :: Maybe Text

The display name for a public registry. This appears on the Amazon ECR Public Gallery.

Only accounts that have the verified account badge can have a registry display name.

newRegistryCatalogData :: RegistryCatalogData

Create a value of RegistryCatalogData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:displayName:RegistryCatalogData', registryCatalogData_displayName - The display name for a public registry. This appears on the Amazon ECR Public Gallery.

Only accounts that have the verified account badge can have a registry display name.

registryCatalogData_displayName :: Lens' RegistryCatalogData (Maybe Text)

The display name for a public registry. This appears on the Amazon ECR Public Gallery.

Only accounts that have the verified account badge can have a registry display name.