Network.AWS.ECRPublic.Types.Repository

Description

Documentation

data Repository

An object representing a repository.

See: newRepository smart constructor.

Constructors

Repository'

Fields

repositoryArn :: Maybe Text

The Amazon Resource Name (ARN) that identifies the repository. The ARN contains the arn:aws:ecr namespace, followed by the region of the repository, AWS account ID of the repository owner, repository namespace, and repository name. For example, arn:aws:ecr:region:012345678910:repository/test.

createdAt :: Maybe POSIX

The date and time, in JavaScript date format, when the repository was created.

registryId :: Maybe Text

The AWS account ID associated with the public registry that contains the repository.

repositoryUri :: Maybe Text

The URI for the repository. You can use this URI for container image push and pull operations.

repositoryName :: Maybe Text

The name of the repository.

newRepository :: Repository

Create a value of Repository with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:repositoryArn:Repository', repository_repositoryArn - The Amazon Resource Name (ARN) that identifies the repository. The ARN contains the arn:aws:ecr namespace, followed by the region of the repository, AWS account ID of the repository owner, repository namespace, and repository name. For example, arn:aws:ecr:region:012345678910:repository/test.

$sel:createdAt:Repository', repository_createdAt - The date and time, in JavaScript date format, when the repository was created.

$sel:registryId:Repository', repository_registryId - The AWS account ID associated with the public registry that contains the repository.

$sel:repositoryUri:Repository', repository_repositoryUri - The URI for the repository. You can use this URI for container image push and pull operations.

$sel:repositoryName:Repository', repository_repositoryName - The name of the repository.

repository_repositoryArn :: Lens' Repository (Maybe Text)

The Amazon Resource Name (ARN) that identifies the repository. The ARN contains the arn:aws:ecr namespace, followed by the region of the repository, AWS account ID of the repository owner, repository namespace, and repository name. For example, arn:aws:ecr:region:012345678910:repository/test.

repository_createdAt :: Lens' Repository (Maybe UTCTime)

The date and time, in JavaScript date format, when the repository was created.

repository_registryId :: Lens' Repository (Maybe Text)

The AWS account ID associated with the public registry that contains the repository.

repository_repositoryUri :: Lens' Repository (Maybe Text)

The URI for the repository. You can use this URI for container image push and pull operations.

repository_repositoryName :: Lens' Repository (Maybe Text)

The name of the repository.