Network.AWS.ECRPublic.Types.RepositoryCatalogDataInput

Description

Documentation

data RepositoryCatalogDataInput

An object containing the catalog data for a repository. This data is publicly visible in the Amazon ECR Public Gallery.

See: newRepositoryCatalogDataInput smart constructor.

Constructors

RepositoryCatalogDataInput'

Fields

logoImageBlob :: Maybe Base64

The base64-encoded repository logo payload.

The repository logo is only publicly visible in the Amazon ECR Public Gallery for verified accounts.

architectures :: Maybe [Text]

The system architecture that the images in the repository are compatible with. On the Amazon ECR Public Gallery, the following supported architectures will appear as badges on the repository and are used as search filters.

Linux

Windows

If an unsupported tag is added to your repository catalog data, it will be associated with the repository and can be retrieved using the API but will not be discoverable in the Amazon ECR Public Gallery.

usageText :: Maybe Text

Detailed information on how to use the contents of the repository. It is publicly visible in the Amazon ECR Public Gallery. The usage text provides context, support information, and additional usage details for users of the repository. The text must be in markdown format.

aboutText :: Maybe Text

A detailed description of the contents of the repository. It is publicly visible in the Amazon ECR Public Gallery. The text must be in markdown format.

operatingSystems :: Maybe [Text]

The operating systems that the images in the repository are compatible with. On the Amazon ECR Public Gallery, the following supported operating systems will appear as badges on the repository and are used as search filters.

ARM

ARM 64

x86

x86-64

If an unsupported tag is added to your repository catalog data, it will be associated with the repository and can be retrieved using the API but will not be discoverable in the Amazon ECR Public Gallery.

description :: Maybe Text

A short description of the contents of the repository. This text appears in both the image details and also when searching for repositories on the Amazon ECR Public Gallery.

newRepositoryCatalogDataInput :: RepositoryCatalogDataInput

Create a value of RepositoryCatalogDataInput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:logoImageBlob:RepositoryCatalogDataInput', repositoryCatalogDataInput_logoImageBlob - The base64-encoded repository logo payload.

The repository logo is only publicly visible in the Amazon ECR Public Gallery for verified accounts.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:architectures:RepositoryCatalogDataInput', repositoryCatalogDataInput_architectures - The system architecture that the images in the repository are compatible with. On the Amazon ECR Public Gallery, the following supported architectures will appear as badges on the repository and are used as search filters.

Linux

Windows

If an unsupported tag is added to your repository catalog data, it will be associated with the repository and can be retrieved using the API but will not be discoverable in the Amazon ECR Public Gallery.

$sel:usageText:RepositoryCatalogDataInput', repositoryCatalogDataInput_usageText - Detailed information on how to use the contents of the repository. It is publicly visible in the Amazon ECR Public Gallery. The usage text provides context, support information, and additional usage details for users of the repository. The text must be in markdown format.

$sel:aboutText:RepositoryCatalogDataInput', repositoryCatalogDataInput_aboutText - A detailed description of the contents of the repository. It is publicly visible in the Amazon ECR Public Gallery. The text must be in markdown format.

$sel:operatingSystems:RepositoryCatalogDataInput', repositoryCatalogDataInput_operatingSystems - The operating systems that the images in the repository are compatible with. On the Amazon ECR Public Gallery, the following supported operating systems will appear as badges on the repository and are used as search filters.

ARM

ARM 64

x86

x86-64

If an unsupported tag is added to your repository catalog data, it will be associated with the repository and can be retrieved using the API but will not be discoverable in the Amazon ECR Public Gallery.

$sel:description:RepositoryCatalogDataInput', repositoryCatalogDataInput_description - A short description of the contents of the repository. This text appears in both the image details and also when searching for repositories on the Amazon ECR Public Gallery.

repositoryCatalogDataInput_logoImageBlob :: Lens' RepositoryCatalogDataInput (Maybe ByteString)

The base64-encoded repository logo payload.

The repository logo is only publicly visible in the Amazon ECR Public Gallery for verified accounts.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

repositoryCatalogDataInput_architectures :: Lens' RepositoryCatalogDataInput (Maybe [Text])

The system architecture that the images in the repository are compatible with. On the Amazon ECR Public Gallery, the following supported architectures will appear as badges on the repository and are used as search filters.

Linux

Windows

If an unsupported tag is added to your repository catalog data, it will be associated with the repository and can be retrieved using the API but will not be discoverable in the Amazon ECR Public Gallery.

repositoryCatalogDataInput_usageText :: Lens' RepositoryCatalogDataInput (Maybe Text)

Detailed information on how to use the contents of the repository. It is publicly visible in the Amazon ECR Public Gallery. The usage text provides context, support information, and additional usage details for users of the repository. The text must be in markdown format.

repositoryCatalogDataInput_aboutText :: Lens' RepositoryCatalogDataInput (Maybe Text)

A detailed description of the contents of the repository. It is publicly visible in the Amazon ECR Public Gallery. The text must be in markdown format.

repositoryCatalogDataInput_operatingSystems :: Lens' RepositoryCatalogDataInput (Maybe [Text])

The operating systems that the images in the repository are compatible with. On the Amazon ECR Public Gallery, the following supported operating systems will appear as badges on the repository and are used as search filters.

ARM

ARM 64

x86

x86-64

If an unsupported tag is added to your repository catalog data, it will be associated with the repository and can be retrieved using the API but will not be discoverable in the Amazon ECR Public Gallery.

repositoryCatalogDataInput_description :: Lens' RepositoryCatalogDataInput (Maybe Text)

A short description of the contents of the repository. This text appears in both the image details and also when searching for repositories on the Amazon ECR Public Gallery.