Network.AWS.ECS.CreateTaskSet

Description

Create a task set in the specified cluster and service. This is used when a service uses the EXTERNAL deployment controller type. For more information, see Amazon ECS Deployment Types in the Amazon Elastic Container Service Developer Guide.

Creating a Request

data CreateTaskSet

See: newCreateTaskSet smart constructor.

Constructors

CreateTaskSet'

Fields

capacityProviderStrategy :: Maybe [CapacityProviderStrategyItem]

The capacity provider strategy to use for the task set.

A capacity provider strategy consists of one or more capacity providers along with the base and weight to assign to them. A capacity provider must be associated with the cluster to be used in a capacity provider strategy. The PutClusterCapacityProviders API is used to associate a capacity provider with a cluster. Only capacity providers with an ACTIVE or UPDATING status can be used.

If a capacityProviderStrategy is specified, the launchType parameter must be omitted. If no capacityProviderStrategy or launchType is specified, the defaultCapacityProviderStrategy for the cluster is used.

If specifying a capacity provider that uses an Auto Scaling group, the capacity provider must already be created. New capacity providers can be created with the CreateCapacityProvider API operation.

To use a Fargate capacity provider, specify either the FARGATE or FARGATE_SPOT capacity providers. The Fargate capacity providers are available to all accounts and only need to be associated with a cluster to be used.

The PutClusterCapacityProviders API operation is used to update the list of available capacity providers for a cluster after the cluster is created.

networkConfiguration :: Maybe NetworkConfiguration

An object representing the network configuration for a task set.

launchType :: Maybe LaunchType

The launch type that new tasks in the task set will use. For more information, see Amazon ECS Launch Types in the Amazon Elastic Container Service Developer Guide.

If a launchType is specified, the capacityProviderStrategy parameter must be omitted.

platformVersion :: Maybe Text

The platform version that the tasks in the task set should use. A platform version is specified only for tasks using the Fargate launch type. If one isn't specified, the LATEST platform version is used by default.

loadBalancers :: Maybe [LoadBalancer]

A load balancer object representing the load balancer to use with the task set. The supported load balancer types are either an Application Load Balancer or a Network Load Balancer.

tags :: Maybe [Tag]

The metadata that you apply to the task set to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define. When a service is deleted, the tags are deleted as well.

The following basic restrictions apply to tags:

Maximum number of tags per resource - 50

For each resource, each tag key must be unique, and each tag key can have only one value.

Maximum key length - 128 Unicode characters in UTF-8

Maximum value length - 256 Unicode characters in UTF-8

If your tagging schema is used across multiple services and resources, remember that other services may have restrictions on allowed characters. Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following characters: + - = . _ : / @.

Tag keys and values are case-sensitive.

Do not use aws:, AWS:, or any upper or lowercase combination of such as a prefix for either keys or values as it is reserved for Amazon Web Services use. You cannot edit or delete tag keys or values with this prefix. Tags with this prefix do not count against your tags per resource limit.

serviceRegistries :: Maybe [ServiceRegistry]

The details of the service discovery registries to assign to this task set. For more information, see Service Discovery.

scale :: Maybe Scale

A floating-point percentage of the desired number of tasks to place and keep running in the task set.

externalId :: Maybe Text

An optional non-unique tag that identifies this task set in external systems. If the task set is associated with a service discovery registry, the tasks in this task set will have the ECS_TASK_SET_EXTERNAL_ID Cloud Map attribute set to the provided value.

clientToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Up to 32 ASCII characters are allowed.

service :: Text

The short name or full Amazon Resource Name (ARN) of the service to create the task set in.

cluster :: Text

The short name or full Amazon Resource Name (ARN) of the cluster that hosts the service to create the task set in.

taskDefinition :: Text

The task definition for the tasks in the task set to use.

newCreateTaskSet

Arguments

:: Text
$sel:service:CreateTaskSet'

-> Text
$sel:cluster:CreateTaskSet'

-> Text
$sel:taskDefinition:CreateTaskSet'

-> CreateTaskSet

Create a value of CreateTaskSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:capacityProviderStrategy:CreateTaskSet', createTaskSet_capacityProviderStrategy - The capacity provider strategy to use for the task set.

A capacity provider strategy consists of one or more capacity providers along with the base and weight to assign to them. A capacity provider must be associated with the cluster to be used in a capacity provider strategy. The PutClusterCapacityProviders API is used to associate a capacity provider with a cluster. Only capacity providers with an ACTIVE or UPDATING status can be used.

If a capacityProviderStrategy is specified, the launchType parameter must be omitted. If no capacityProviderStrategy or launchType is specified, the defaultCapacityProviderStrategy for the cluster is used.

If specifying a capacity provider that uses an Auto Scaling group, the capacity provider must already be created. New capacity providers can be created with the CreateCapacityProvider API operation.

To use a Fargate capacity provider, specify either the FARGATE or FARGATE_SPOT capacity providers. The Fargate capacity providers are available to all accounts and only need to be associated with a cluster to be used.

The PutClusterCapacityProviders API operation is used to update the list of available capacity providers for a cluster after the cluster is created.

$sel:networkConfiguration:CreateTaskSet', createTaskSet_networkConfiguration - An object representing the network configuration for a task set.

$sel:launchType:CreateTaskSet', createTaskSet_launchType - The launch type that new tasks in the task set will use. For more information, see Amazon ECS Launch Types in the Amazon Elastic Container Service Developer Guide.

If a launchType is specified, the capacityProviderStrategy parameter must be omitted.

$sel:platformVersion:CreateTaskSet', createTaskSet_platformVersion - The platform version that the tasks in the task set should use. A platform version is specified only for tasks using the Fargate launch type. If one isn't specified, the LATEST platform version is used by default.

$sel:loadBalancers:CreateTaskSet', createTaskSet_loadBalancers - A load balancer object representing the load balancer to use with the task set. The supported load balancer types are either an Application Load Balancer or a Network Load Balancer.

$sel:tags:CreateTaskSet', createTaskSet_tags - The metadata that you apply to the task set to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define. When a service is deleted, the tags are deleted as well.

The following basic restrictions apply to tags:

Maximum number of tags per resource - 50

For each resource, each tag key must be unique, and each tag key can have only one value.

Maximum key length - 128 Unicode characters in UTF-8

Maximum value length - 256 Unicode characters in UTF-8

If your tagging schema is used across multiple services and resources, remember that other services may have restrictions on allowed characters. Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following characters: + - = . _ : / @.

Tag keys and values are case-sensitive.

Do not use aws:, AWS:, or any upper or lowercase combination of such as a prefix for either keys or values as it is reserved for Amazon Web Services use. You cannot edit or delete tag keys or values with this prefix. Tags with this prefix do not count against your tags per resource limit.

$sel:serviceRegistries:CreateTaskSet', createTaskSet_serviceRegistries - The details of the service discovery registries to assign to this task set. For more information, see Service Discovery.

$sel:scale:CreateTaskSet', createTaskSet_scale - A floating-point percentage of the desired number of tasks to place and keep running in the task set.

$sel:externalId:CreateTaskSet', createTaskSet_externalId - An optional non-unique tag that identifies this task set in external systems. If the task set is associated with a service discovery registry, the tasks in this task set will have the ECS_TASK_SET_EXTERNAL_ID Cloud Map attribute set to the provided value.

$sel:clientToken:CreateTaskSet', createTaskSet_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Up to 32 ASCII characters are allowed.

$sel:service:CreateTaskSet', createTaskSet_service - The short name or full Amazon Resource Name (ARN) of the service to create the task set in.

$sel:cluster:CreateTaskSet', createTaskSet_cluster - The short name or full Amazon Resource Name (ARN) of the cluster that hosts the service to create the task set in.

$sel:taskDefinition:CreateTaskSet', createTaskSet_taskDefinition - The task definition for the tasks in the task set to use.

Request Lenses

createTaskSet_capacityProviderStrategy :: Lens' CreateTaskSet (Maybe [CapacityProviderStrategyItem])

The capacity provider strategy to use for the task set.

A capacity provider strategy consists of one or more capacity providers along with the base and weight to assign to them. A capacity provider must be associated with the cluster to be used in a capacity provider strategy. The PutClusterCapacityProviders API is used to associate a capacity provider with a cluster. Only capacity providers with an ACTIVE or UPDATING status can be used.

If a capacityProviderStrategy is specified, the launchType parameter must be omitted. If no capacityProviderStrategy or launchType is specified, the defaultCapacityProviderStrategy for the cluster is used.

If specifying a capacity provider that uses an Auto Scaling group, the capacity provider must already be created. New capacity providers can be created with the CreateCapacityProvider API operation.

To use a Fargate capacity provider, specify either the FARGATE or FARGATE_SPOT capacity providers. The Fargate capacity providers are available to all accounts and only need to be associated with a cluster to be used.

The PutClusterCapacityProviders API operation is used to update the list of available capacity providers for a cluster after the cluster is created.

createTaskSet_networkConfiguration :: Lens' CreateTaskSet (Maybe NetworkConfiguration)

An object representing the network configuration for a task set.

createTaskSet_launchType :: Lens' CreateTaskSet (Maybe LaunchType)

The launch type that new tasks in the task set will use. For more information, see Amazon ECS Launch Types in the Amazon Elastic Container Service Developer Guide.

If a launchType is specified, the capacityProviderStrategy parameter must be omitted.

createTaskSet_platformVersion :: Lens' CreateTaskSet (Maybe Text)

The platform version that the tasks in the task set should use. A platform version is specified only for tasks using the Fargate launch type. If one isn't specified, the LATEST platform version is used by default.

createTaskSet_loadBalancers :: Lens' CreateTaskSet (Maybe [LoadBalancer])

A load balancer object representing the load balancer to use with the task set. The supported load balancer types are either an Application Load Balancer or a Network Load Balancer.

createTaskSet_tags :: Lens' CreateTaskSet (Maybe [Tag])

The metadata that you apply to the task set to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define. When a service is deleted, the tags are deleted as well.

The following basic restrictions apply to tags:

Maximum number of tags per resource - 50

For each resource, each tag key must be unique, and each tag key can have only one value.

Maximum key length - 128 Unicode characters in UTF-8

Maximum value length - 256 Unicode characters in UTF-8

If your tagging schema is used across multiple services and resources, remember that other services may have restrictions on allowed characters. Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following characters: + - = . _ : / @.

Tag keys and values are case-sensitive.

Do not use aws:, AWS:, or any upper or lowercase combination of such as a prefix for either keys or values as it is reserved for Amazon Web Services use. You cannot edit or delete tag keys or values with this prefix. Tags with this prefix do not count against your tags per resource limit.

createTaskSet_serviceRegistries :: Lens' CreateTaskSet (Maybe [ServiceRegistry])

The details of the service discovery registries to assign to this task set. For more information, see Service Discovery.

createTaskSet_scale :: Lens' CreateTaskSet (Maybe Scale)

A floating-point percentage of the desired number of tasks to place and keep running in the task set.

createTaskSet_externalId :: Lens' CreateTaskSet (Maybe Text)

An optional non-unique tag that identifies this task set in external systems. If the task set is associated with a service discovery registry, the tasks in this task set will have the ECS_TASK_SET_EXTERNAL_ID Cloud Map attribute set to the provided value.

createTaskSet_clientToken :: Lens' CreateTaskSet (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Up to 32 ASCII characters are allowed.

createTaskSet_service :: Lens' CreateTaskSet Text

The short name or full Amazon Resource Name (ARN) of the service to create the task set in.

createTaskSet_cluster :: Lens' CreateTaskSet Text

The short name or full Amazon Resource Name (ARN) of the cluster that hosts the service to create the task set in.

createTaskSet_taskDefinition :: Lens' CreateTaskSet Text

The task definition for the tasks in the task set to use.

Destructuring the Response

data CreateTaskSetResponse

See: newCreateTaskSetResponse smart constructor.

Constructors

CreateTaskSetResponse'

Fields

taskSet :: Maybe TaskSet

Information about a set of Amazon ECS tasks in either an CodeDeploy or an EXTERNAL deployment. A task set includes details such as the desired number of tasks, how many tasks are running, and whether the task set serves production traffic.

httpStatus :: Int

The response's http status code.

newCreateTaskSetResponse

Arguments

:: Int
$sel:httpStatus:CreateTaskSetResponse'

-> CreateTaskSetResponse

Create a value of CreateTaskSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:taskSet:CreateTaskSetResponse', createTaskSetResponse_taskSet - Information about a set of Amazon ECS tasks in either an CodeDeploy or an EXTERNAL deployment. A task set includes details such as the desired number of tasks, how many tasks are running, and whether the task set serves production traffic.

$sel:httpStatus:CreateTaskSetResponse', createTaskSetResponse_httpStatus - The response's http status code.

Response Lenses

createTaskSetResponse_taskSet :: Lens' CreateTaskSetResponse (Maybe TaskSet)

Information about a set of Amazon ECS tasks in either an CodeDeploy or an EXTERNAL deployment. A task set includes details such as the desired number of tasks, how many tasks are running, and whether the task set serves production traffic.

createTaskSetResponse_httpStatus :: Lens' CreateTaskSetResponse Int

The response's http status code.