Network.AWS.ECS.DeleteAccountSetting

Description

Disables an account setting for a specified IAM user, IAM role, or the root user for an account.

Creating a Request

data DeleteAccountSetting

See: newDeleteAccountSetting smart constructor.

Constructors

DeleteAccountSetting'

Fields

principalArn :: Maybe Text

The ARN of the principal, which can be an IAM user, IAM role, or the root user. If you specify the root user, it disables the account setting for all IAM users, IAM roles, and the root user of the account unless an IAM user or role explicitly overrides these settings. If this field is omitted, the setting is changed only for the authenticated user.

name :: SettingName

The resource name for which to disable the account setting. If serviceLongArnFormat is specified, the ARN for your Amazon ECS services is affected. If taskLongArnFormat is specified, the ARN and resource ID for your Amazon ECS tasks is affected. If containerInstanceLongArnFormat is specified, the ARN and resource ID for your Amazon ECS container instances is affected. If awsvpcTrunking is specified, the ENI limit for your Amazon ECS container instances is affected.

newDeleteAccountSetting

Arguments

:: SettingName
$sel:name:DeleteAccountSetting'

-> DeleteAccountSetting

Create a value of DeleteAccountSetting with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:principalArn:DeleteAccountSetting', deleteAccountSetting_principalArn - The ARN of the principal, which can be an IAM user, IAM role, or the root user. If you specify the root user, it disables the account setting for all IAM users, IAM roles, and the root user of the account unless an IAM user or role explicitly overrides these settings. If this field is omitted, the setting is changed only for the authenticated user.

$sel:name:DeleteAccountSetting', deleteAccountSetting_name - The resource name for which to disable the account setting. If serviceLongArnFormat is specified, the ARN for your Amazon ECS services is affected. If taskLongArnFormat is specified, the ARN and resource ID for your Amazon ECS tasks is affected. If containerInstanceLongArnFormat is specified, the ARN and resource ID for your Amazon ECS container instances is affected. If awsvpcTrunking is specified, the ENI limit for your Amazon ECS container instances is affected.

Request Lenses

deleteAccountSetting_principalArn :: Lens' DeleteAccountSetting (Maybe Text)

The ARN of the principal, which can be an IAM user, IAM role, or the root user. If you specify the root user, it disables the account setting for all IAM users, IAM roles, and the root user of the account unless an IAM user or role explicitly overrides these settings. If this field is omitted, the setting is changed only for the authenticated user.

deleteAccountSetting_name :: Lens' DeleteAccountSetting SettingName

The resource name for which to disable the account setting. If serviceLongArnFormat is specified, the ARN for your Amazon ECS services is affected. If taskLongArnFormat is specified, the ARN and resource ID for your Amazon ECS tasks is affected. If containerInstanceLongArnFormat is specified, the ARN and resource ID for your Amazon ECS container instances is affected. If awsvpcTrunking is specified, the ENI limit for your Amazon ECS container instances is affected.

Destructuring the Response

data DeleteAccountSettingResponse

See: newDeleteAccountSettingResponse smart constructor.

Constructors

DeleteAccountSettingResponse'

Fields

setting :: Maybe Setting

The account setting for the specified principal ARN.

httpStatus :: Int

The response's http status code.

newDeleteAccountSettingResponse

Arguments

:: Int
$sel:httpStatus:DeleteAccountSettingResponse'

-> DeleteAccountSettingResponse

Create a value of DeleteAccountSettingResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:setting:DeleteAccountSettingResponse', deleteAccountSettingResponse_setting - The account setting for the specified principal ARN.

$sel:httpStatus:DeleteAccountSettingResponse', deleteAccountSettingResponse_httpStatus - The response's http status code.

Response Lenses

deleteAccountSettingResponse_setting :: Lens' DeleteAccountSettingResponse (Maybe Setting)

The account setting for the specified principal ARN.

deleteAccountSettingResponse_httpStatus :: Lens' DeleteAccountSettingResponse Int

The response's http status code.