Network.AWS.ECS.DescribeCapacityProviders

Description

Describes one or more of your capacity providers.

Creating a Request

data DescribeCapacityProviders

See: newDescribeCapacityProviders smart constructor.

Constructors

DescribeCapacityProviders'

Fields

include :: Maybe [CapacityProviderField]

Specifies whether or not you want to see the resource tags for the capacity provider. If TAGS is specified, the tags are included in the response. If this field is omitted, tags are not included in the response.

nextToken :: Maybe Text

The nextToken value returned from a previous paginated DescribeCapacityProviders request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value.

This token should be treated as an opaque identifier that is only used to retrieve the next items in a list and not for other programmatic purposes.

capacityProviders :: Maybe [Text]

The short name or full Amazon Resource Name (ARN) of one or more capacity providers. Up to 100 capacity providers can be described in an action.

maxResults :: Maybe Int

The maximum number of account setting results returned by DescribeCapacityProviders in paginated output. When this parameter is used, DescribeCapacityProviders only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another DescribeCapacityProviders request with the returned nextToken value. This value can be between 1 and 10. If this parameter is not used, then DescribeCapacityProviders returns up to 10 results and a nextToken value if applicable.

newDescribeCapacityProviders :: DescribeCapacityProviders

Create a value of DescribeCapacityProviders with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:include:DescribeCapacityProviders', describeCapacityProviders_include - Specifies whether or not you want to see the resource tags for the capacity provider. If TAGS is specified, the tags are included in the response. If this field is omitted, tags are not included in the response.

$sel:nextToken:DescribeCapacityProviders', describeCapacityProviders_nextToken - The nextToken value returned from a previous paginated DescribeCapacityProviders request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value.

This token should be treated as an opaque identifier that is only used to retrieve the next items in a list and not for other programmatic purposes.

$sel:capacityProviders:DescribeCapacityProviders', describeCapacityProviders_capacityProviders - The short name or full Amazon Resource Name (ARN) of one or more capacity providers. Up to 100 capacity providers can be described in an action.

$sel:maxResults:DescribeCapacityProviders', describeCapacityProviders_maxResults - The maximum number of account setting results returned by DescribeCapacityProviders in paginated output. When this parameter is used, DescribeCapacityProviders only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another DescribeCapacityProviders request with the returned nextToken value. This value can be between 1 and 10. If this parameter is not used, then DescribeCapacityProviders returns up to 10 results and a nextToken value if applicable.

Request Lenses

describeCapacityProviders_include :: Lens' DescribeCapacityProviders (Maybe [CapacityProviderField])

Specifies whether or not you want to see the resource tags for the capacity provider. If TAGS is specified, the tags are included in the response. If this field is omitted, tags are not included in the response.

describeCapacityProviders_nextToken :: Lens' DescribeCapacityProviders (Maybe Text)

The nextToken value returned from a previous paginated DescribeCapacityProviders request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value.

This token should be treated as an opaque identifier that is only used to retrieve the next items in a list and not for other programmatic purposes.

describeCapacityProviders_capacityProviders :: Lens' DescribeCapacityProviders (Maybe [Text])

The short name or full Amazon Resource Name (ARN) of one or more capacity providers. Up to 100 capacity providers can be described in an action.

describeCapacityProviders_maxResults :: Lens' DescribeCapacityProviders (Maybe Int)

The maximum number of account setting results returned by DescribeCapacityProviders in paginated output. When this parameter is used, DescribeCapacityProviders only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another DescribeCapacityProviders request with the returned nextToken value. This value can be between 1 and 10. If this parameter is not used, then DescribeCapacityProviders returns up to 10 results and a nextToken value if applicable.

Destructuring the Response

data DescribeCapacityProvidersResponse

See: newDescribeCapacityProvidersResponse smart constructor.

Constructors

DescribeCapacityProvidersResponse'

Fields

failures :: Maybe [Failure]

Any failures associated with the call.

nextToken :: Maybe Text

The nextToken value to include in a future DescribeCapacityProviders request. When the results of a DescribeCapacityProviders request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

capacityProviders :: Maybe [CapacityProvider]

The list of capacity providers.

httpStatus :: Int

The response's http status code.

newDescribeCapacityProvidersResponse

Arguments

:: Int
$sel:httpStatus:DescribeCapacityProvidersResponse'

-> DescribeCapacityProvidersResponse

Create a value of DescribeCapacityProvidersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failures:DescribeCapacityProvidersResponse', describeCapacityProvidersResponse_failures - Any failures associated with the call.

$sel:nextToken:DescribeCapacityProviders', describeCapacityProvidersResponse_nextToken - The nextToken value to include in a future DescribeCapacityProviders request. When the results of a DescribeCapacityProviders request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

$sel:capacityProviders:DescribeCapacityProviders', describeCapacityProvidersResponse_capacityProviders - The list of capacity providers.

$sel:httpStatus:DescribeCapacityProvidersResponse', describeCapacityProvidersResponse_httpStatus - The response's http status code.

Response Lenses

describeCapacityProvidersResponse_failures :: Lens' DescribeCapacityProvidersResponse (Maybe [Failure])

Any failures associated with the call.

describeCapacityProvidersResponse_nextToken :: Lens' DescribeCapacityProvidersResponse (Maybe Text)

The nextToken value to include in a future DescribeCapacityProviders request. When the results of a DescribeCapacityProviders request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

describeCapacityProvidersResponse_capacityProviders :: Lens' DescribeCapacityProvidersResponse (Maybe [CapacityProvider])

The list of capacity providers.

describeCapacityProvidersResponse_httpStatus :: Lens' DescribeCapacityProvidersResponse Int

The response's http status code.