Network.AWS.ECS.DescribeClusters

Description

Describes one or more of your clusters.

Creating a Request

data DescribeClusters

See: newDescribeClusters smart constructor.

Constructors

DescribeClusters'

Fields

include :: Maybe [ClusterField]

Whether to include additional information about the clusters in the response. If this field is omitted, this information isn't included.

If ATTACHMENTS is specified, the attachments for the container instances or tasks within the cluster are included.

If SETTINGS is specified, the settings for the cluster are included.

If CONFIGURATIONS is specified, the configuration for the cluster is included.

If STATISTICS is specified, the task and service count is included, separated by launch type.

If TAGS is specified, the metadata tags associated with the cluster are included.

clusters :: Maybe [Text]

A list of up to 100 cluster names or full cluster Amazon Resource Name (ARN) entries. If you do not specify a cluster, the default cluster is assumed.

newDescribeClusters :: DescribeClusters

Create a value of DescribeClusters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:include:DescribeClusters', describeClusters_include - Whether to include additional information about the clusters in the response. If this field is omitted, this information isn't included.

If ATTACHMENTS is specified, the attachments for the container instances or tasks within the cluster are included.

If SETTINGS is specified, the settings for the cluster are included.

If CONFIGURATIONS is specified, the configuration for the cluster is included.

If STATISTICS is specified, the task and service count is included, separated by launch type.

If TAGS is specified, the metadata tags associated with the cluster are included.

$sel:clusters:DescribeClusters', describeClusters_clusters - A list of up to 100 cluster names or full cluster Amazon Resource Name (ARN) entries. If you do not specify a cluster, the default cluster is assumed.

Request Lenses

describeClusters_include :: Lens' DescribeClusters (Maybe [ClusterField])

Whether to include additional information about the clusters in the response. If this field is omitted, this information isn't included.

If ATTACHMENTS is specified, the attachments for the container instances or tasks within the cluster are included.

If SETTINGS is specified, the settings for the cluster are included.

If CONFIGURATIONS is specified, the configuration for the cluster is included.

If STATISTICS is specified, the task and service count is included, separated by launch type.

If TAGS is specified, the metadata tags associated with the cluster are included.

describeClusters_clusters :: Lens' DescribeClusters (Maybe [Text])

A list of up to 100 cluster names or full cluster Amazon Resource Name (ARN) entries. If you do not specify a cluster, the default cluster is assumed.

Destructuring the Response

data DescribeClustersResponse

See: newDescribeClustersResponse smart constructor.

Constructors

DescribeClustersResponse'

Fields

failures :: Maybe [Failure]

Any failures associated with the call.

clusters :: Maybe [Cluster]

The list of clusters.

httpStatus :: Int

The response's http status code.

newDescribeClustersResponse

Arguments

:: Int
$sel:httpStatus:DescribeClustersResponse'

-> DescribeClustersResponse

Create a value of DescribeClustersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failures:DescribeClustersResponse', describeClustersResponse_failures - Any failures associated with the call.

$sel:clusters:DescribeClusters', describeClustersResponse_clusters - The list of clusters.

$sel:httpStatus:DescribeClustersResponse', describeClustersResponse_httpStatus - The response's http status code.

Response Lenses

describeClustersResponse_failures :: Lens' DescribeClustersResponse (Maybe [Failure])

Any failures associated with the call.

describeClustersResponse_clusters :: Lens' DescribeClustersResponse (Maybe [Cluster])

The list of clusters.

describeClustersResponse_httpStatus :: Lens' DescribeClustersResponse Int

The response's http status code.