Network.AWS.ECS.DescribeContainerInstances

Description

Describes one or more container instances. Returns metadata about each container instance requested.

Creating a Request

data DescribeContainerInstances

See: newDescribeContainerInstances smart constructor.

Constructors

DescribeContainerInstances'

Fields

include :: Maybe [ContainerInstanceField]

Specifies whether you want to see the resource tags for the container instance. If TAGS is specified, the tags are included in the response. If this field is omitted, tags are not included in the response.

cluster :: Maybe Text

The short name or full Amazon Resource Name (ARN) of the cluster that hosts the container instances to describe. If you do not specify a cluster, the default cluster is assumed. This parameter is required if the container instance or container instances you are describing were launched in any cluster other than the default cluster.

containerInstances :: [Text]

A list of up to 100 container instance IDs or full Amazon Resource Name (ARN) entries.

newDescribeContainerInstances :: DescribeContainerInstances

Create a value of DescribeContainerInstances with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:include:DescribeContainerInstances', describeContainerInstances_include - Specifies whether you want to see the resource tags for the container instance. If TAGS is specified, the tags are included in the response. If this field is omitted, tags are not included in the response.

$sel:cluster:DescribeContainerInstances', describeContainerInstances_cluster - The short name or full Amazon Resource Name (ARN) of the cluster that hosts the container instances to describe. If you do not specify a cluster, the default cluster is assumed. This parameter is required if the container instance or container instances you are describing were launched in any cluster other than the default cluster.

$sel:containerInstances:DescribeContainerInstances', describeContainerInstances_containerInstances - A list of up to 100 container instance IDs or full Amazon Resource Name (ARN) entries.

Request Lenses

describeContainerInstances_include :: Lens' DescribeContainerInstances (Maybe [ContainerInstanceField])

Specifies whether you want to see the resource tags for the container instance. If TAGS is specified, the tags are included in the response. If this field is omitted, tags are not included in the response.

describeContainerInstances_cluster :: Lens' DescribeContainerInstances (Maybe Text)

The short name or full Amazon Resource Name (ARN) of the cluster that hosts the container instances to describe. If you do not specify a cluster, the default cluster is assumed. This parameter is required if the container instance or container instances you are describing were launched in any cluster other than the default cluster.

describeContainerInstances_containerInstances :: Lens' DescribeContainerInstances [Text]

A list of up to 100 container instance IDs or full Amazon Resource Name (ARN) entries.

Destructuring the Response

data DescribeContainerInstancesResponse

See: newDescribeContainerInstancesResponse smart constructor.

Constructors

DescribeContainerInstancesResponse'

Fields

failures :: Maybe [Failure]

Any failures associated with the call.

containerInstances :: Maybe [ContainerInstance]

The list of container instances.

httpStatus :: Int

The response's http status code.

newDescribeContainerInstancesResponse

Arguments

:: Int
$sel:httpStatus:DescribeContainerInstancesResponse'

-> DescribeContainerInstancesResponse

Create a value of DescribeContainerInstancesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failures:DescribeContainerInstancesResponse', describeContainerInstancesResponse_failures - Any failures associated with the call.

$sel:containerInstances:DescribeContainerInstances', describeContainerInstancesResponse_containerInstances - The list of container instances.

$sel:httpStatus:DescribeContainerInstancesResponse', describeContainerInstancesResponse_httpStatus - The response's http status code.

Response Lenses

describeContainerInstancesResponse_failures :: Lens' DescribeContainerInstancesResponse (Maybe [Failure])

Any failures associated with the call.

describeContainerInstancesResponse_containerInstances :: Lens' DescribeContainerInstancesResponse (Maybe [ContainerInstance])

The list of container instances.

describeContainerInstancesResponse_httpStatus :: Lens' DescribeContainerInstancesResponse Int

The response's http status code.