Network.AWS.ECS.DescribeTaskDefinition

Description

Describes a task definition. You can specify a family and revision to find information about a specific task definition, or you can simply specify the family to find the latest ACTIVE revision in that family.

You can only describe INACTIVE task definitions while an active task or service references them.

Creating a Request

data DescribeTaskDefinition

See: newDescribeTaskDefinition smart constructor.

Constructors

DescribeTaskDefinition'

Fields

include :: Maybe [TaskDefinitionField]

Specifies whether to see the resource tags for the task definition. If TAGS is specified, the tags are included in the response. If this field is omitted, tags are not included in the response.

taskDefinition :: Text

The family for the latest ACTIVE revision, family and revision (family:revision) for a specific revision in the family, or full Amazon Resource Name (ARN) of the task definition to describe.

newDescribeTaskDefinition

Arguments

:: Text
$sel:taskDefinition:DescribeTaskDefinition'

-> DescribeTaskDefinition

Create a value of DescribeTaskDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:include:DescribeTaskDefinition', describeTaskDefinition_include - Specifies whether to see the resource tags for the task definition. If TAGS is specified, the tags are included in the response. If this field is omitted, tags are not included in the response.

$sel:taskDefinition:DescribeTaskDefinition', describeTaskDefinition_taskDefinition - The family for the latest ACTIVE revision, family and revision (family:revision) for a specific revision in the family, or full Amazon Resource Name (ARN) of the task definition to describe.

Request Lenses

describeTaskDefinition_include :: Lens' DescribeTaskDefinition (Maybe [TaskDefinitionField])

Specifies whether to see the resource tags for the task definition. If TAGS is specified, the tags are included in the response. If this field is omitted, tags are not included in the response.

describeTaskDefinition_taskDefinition :: Lens' DescribeTaskDefinition Text

The family for the latest ACTIVE revision, family and revision (family:revision) for a specific revision in the family, or full Amazon Resource Name (ARN) of the task definition to describe.

Destructuring the Response

data DescribeTaskDefinitionResponse

See: newDescribeTaskDefinitionResponse smart constructor.

Constructors

DescribeTaskDefinitionResponse'

Fields

taskDefinition :: Maybe TaskDefinition

The full task definition description.

tags :: Maybe [Tag]

The metadata that is applied to the task definition to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define.

The following basic restrictions apply to tags:

Maximum number of tags per resource - 50

For each resource, each tag key must be unique, and each tag key can have only one value.

Maximum key length - 128 Unicode characters in UTF-8

Maximum value length - 256 Unicode characters in UTF-8

If your tagging schema is used across multiple services and resources, remember that other services may have restrictions on allowed characters. Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following characters: + - = . _ : / @.

Tag keys and values are case-sensitive.

Do not use aws:, AWS:, or any upper or lowercase combination of such as a prefix for either keys or values as it is reserved for Amazon Web Services use. You cannot edit or delete tag keys or values with this prefix. Tags with this prefix do not count against your tags per resource limit.

httpStatus :: Int

The response's http status code.

newDescribeTaskDefinitionResponse

Arguments

:: Int
$sel:httpStatus:DescribeTaskDefinitionResponse'

-> DescribeTaskDefinitionResponse

Create a value of DescribeTaskDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:taskDefinition:DescribeTaskDefinition', describeTaskDefinitionResponse_taskDefinition - The full task definition description.

$sel:tags:DescribeTaskDefinitionResponse', describeTaskDefinitionResponse_tags - The metadata that is applied to the task definition to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define.

The following basic restrictions apply to tags:

Maximum number of tags per resource - 50

For each resource, each tag key must be unique, and each tag key can have only one value.

Maximum key length - 128 Unicode characters in UTF-8

Maximum value length - 256 Unicode characters in UTF-8

If your tagging schema is used across multiple services and resources, remember that other services may have restrictions on allowed characters. Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following characters: + - = . _ : / @.

Tag keys and values are case-sensitive.

Do not use aws:, AWS:, or any upper or lowercase combination of such as a prefix for either keys or values as it is reserved for Amazon Web Services use. You cannot edit or delete tag keys or values with this prefix. Tags with this prefix do not count against your tags per resource limit.

$sel:httpStatus:DescribeTaskDefinitionResponse', describeTaskDefinitionResponse_httpStatus - The response's http status code.

Response Lenses

describeTaskDefinitionResponse_taskDefinition :: Lens' DescribeTaskDefinitionResponse (Maybe TaskDefinition)

The full task definition description.

describeTaskDefinitionResponse_tags :: Lens' DescribeTaskDefinitionResponse (Maybe [Tag])

The metadata that is applied to the task definition to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define.

The following basic restrictions apply to tags:

Maximum number of tags per resource - 50

For each resource, each tag key must be unique, and each tag key can have only one value.

Maximum key length - 128 Unicode characters in UTF-8

Maximum value length - 256 Unicode characters in UTF-8

If your tagging schema is used across multiple services and resources, remember that other services may have restrictions on allowed characters. Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following characters: + - = . _ : / @.

Tag keys and values are case-sensitive.

Do not use aws:, AWS:, or any upper or lowercase combination of such as a prefix for either keys or values as it is reserved for Amazon Web Services use. You cannot edit or delete tag keys or values with this prefix. Tags with this prefix do not count against your tags per resource limit.

describeTaskDefinitionResponse_httpStatus :: Lens' DescribeTaskDefinitionResponse Int

The response's http status code.