Network.AWS.ECS.DescribeTasks

Description

Describes a specified task or tasks.

Creating a Request

data DescribeTasks

See: newDescribeTasks smart constructor.

Constructors

DescribeTasks'

Fields

include :: Maybe [TaskField]

Specifies whether you want to see the resource tags for the task. If TAGS is specified, the tags are included in the response. If this field is omitted, tags are not included in the response.

cluster :: Maybe Text

The short name or full Amazon Resource Name (ARN) of the cluster that hosts the task or tasks to describe. If you do not specify a cluster, the default cluster is assumed. This parameter is required if the task or tasks you are describing were launched in any cluster other than the default cluster.

tasks :: [Text]

A list of up to 100 task IDs or full ARN entries.

newDescribeTasks :: DescribeTasks

Create a value of DescribeTasks with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:include:DescribeTasks', describeTasks_include - Specifies whether you want to see the resource tags for the task. If TAGS is specified, the tags are included in the response. If this field is omitted, tags are not included in the response.

$sel:cluster:DescribeTasks', describeTasks_cluster - The short name or full Amazon Resource Name (ARN) of the cluster that hosts the task or tasks to describe. If you do not specify a cluster, the default cluster is assumed. This parameter is required if the task or tasks you are describing were launched in any cluster other than the default cluster.

$sel:tasks:DescribeTasks', describeTasks_tasks - A list of up to 100 task IDs or full ARN entries.

Request Lenses

describeTasks_include :: Lens' DescribeTasks (Maybe [TaskField])

Specifies whether you want to see the resource tags for the task. If TAGS is specified, the tags are included in the response. If this field is omitted, tags are not included in the response.

describeTasks_cluster :: Lens' DescribeTasks (Maybe Text)

The short name or full Amazon Resource Name (ARN) of the cluster that hosts the task or tasks to describe. If you do not specify a cluster, the default cluster is assumed. This parameter is required if the task or tasks you are describing were launched in any cluster other than the default cluster.

describeTasks_tasks :: Lens' DescribeTasks [Text]

A list of up to 100 task IDs or full ARN entries.

Destructuring the Response

data DescribeTasksResponse

See: newDescribeTasksResponse smart constructor.

Constructors

DescribeTasksResponse'

Fields

failures :: Maybe [Failure]

Any failures associated with the call.

tasks :: Maybe [Task]

The list of tasks.

httpStatus :: Int

The response's http status code.

newDescribeTasksResponse

Arguments

:: Int
$sel:httpStatus:DescribeTasksResponse'

-> DescribeTasksResponse

Create a value of DescribeTasksResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failures:DescribeTasksResponse', describeTasksResponse_failures - Any failures associated with the call.

$sel:tasks:DescribeTasks', describeTasksResponse_tasks - The list of tasks.

$sel:httpStatus:DescribeTasksResponse', describeTasksResponse_httpStatus - The response's http status code.

Response Lenses

describeTasksResponse_failures :: Lens' DescribeTasksResponse (Maybe [Failure])

Any failures associated with the call.

describeTasksResponse_tasks :: Lens' DescribeTasksResponse (Maybe [Task])

The list of tasks.

describeTasksResponse_httpStatus :: Lens' DescribeTasksResponse Int

The response's http status code.