Network.AWS.ECS.ExecuteCommand

Description

Runs a command remotely on a container within a task.

Creating a Request

data ExecuteCommand

See: newExecuteCommand smart constructor.

Constructors

ExecuteCommand'

Fields

container :: Maybe Text

The name of the container to execute the command on. A container name only needs to be specified for tasks containing multiple containers.

cluster :: Maybe Text

The Amazon Resource Name (ARN) or short name of the cluster the task is running in. If you do not specify a cluster, the default cluster is assumed.

command :: Text

The command to run on the container.

interactive :: Bool

Use this flag to run your command in interactive mode.

task :: Text

The Amazon Resource Name (ARN) or ID of the task the container is part of.

newExecuteCommand

Arguments

:: Text
$sel:command:ExecuteCommand'

-> Bool
$sel:interactive:ExecuteCommand'

-> Text
$sel:task:ExecuteCommand'

-> ExecuteCommand

Create a value of ExecuteCommand with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:container:ExecuteCommand', executeCommand_container - The name of the container to execute the command on. A container name only needs to be specified for tasks containing multiple containers.

$sel:cluster:ExecuteCommand', executeCommand_cluster - The Amazon Resource Name (ARN) or short name of the cluster the task is running in. If you do not specify a cluster, the default cluster is assumed.

$sel:command:ExecuteCommand', executeCommand_command - The command to run on the container.

$sel:interactive:ExecuteCommand', executeCommand_interactive - Use this flag to run your command in interactive mode.

$sel:task:ExecuteCommand', executeCommand_task - The Amazon Resource Name (ARN) or ID of the task the container is part of.

Request Lenses

executeCommand_container :: Lens' ExecuteCommand (Maybe Text)

The name of the container to execute the command on. A container name only needs to be specified for tasks containing multiple containers.

executeCommand_cluster :: Lens' ExecuteCommand (Maybe Text)

The Amazon Resource Name (ARN) or short name of the cluster the task is running in. If you do not specify a cluster, the default cluster is assumed.

executeCommand_command :: Lens' ExecuteCommand Text

The command to run on the container.

executeCommand_interactive :: Lens' ExecuteCommand Bool

Use this flag to run your command in interactive mode.

executeCommand_task :: Lens' ExecuteCommand Text

The Amazon Resource Name (ARN) or ID of the task the container is part of.

Destructuring the Response

data ExecuteCommandResponse

See: newExecuteCommandResponse smart constructor.

Constructors

ExecuteCommandResponse'

Fields

clusterArn :: Maybe Text

The Amazon Resource Name (ARN) of the cluster.

interactive :: Maybe Bool

Whether or not the execute command session is running in interactive mode. Amazon ECS only supports initiating interactive sessions, so you must specify true for this value.

containerArn :: Maybe Text

The Amazon Resource Name (ARN) of the container.

session :: Maybe Session

The details of the SSM session that was created for this instance of execute-command.

containerName :: Maybe Text

The name of the container.

taskArn :: Maybe Text

The Amazon Resource Name (ARN) of the task.

httpStatus :: Int

The response's http status code.

newExecuteCommandResponse

Arguments

:: Int
$sel:httpStatus:ExecuteCommandResponse'

-> ExecuteCommandResponse

Create a value of ExecuteCommandResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clusterArn:ExecuteCommandResponse', executeCommandResponse_clusterArn - The Amazon Resource Name (ARN) of the cluster.

$sel:interactive:ExecuteCommand', executeCommandResponse_interactive - Whether or not the execute command session is running in interactive mode. Amazon ECS only supports initiating interactive sessions, so you must specify true for this value.

$sel:containerArn:ExecuteCommandResponse', executeCommandResponse_containerArn - The Amazon Resource Name (ARN) of the container.

$sel:session:ExecuteCommandResponse', executeCommandResponse_session - The details of the SSM session that was created for this instance of execute-command.

$sel:containerName:ExecuteCommandResponse', executeCommandResponse_containerName - The name of the container.

$sel:taskArn:ExecuteCommandResponse', executeCommandResponse_taskArn - The Amazon Resource Name (ARN) of the task.

$sel:httpStatus:ExecuteCommandResponse', executeCommandResponse_httpStatus - The response's http status code.

Response Lenses

executeCommandResponse_clusterArn :: Lens' ExecuteCommandResponse (Maybe Text)

The Amazon Resource Name (ARN) of the cluster.

executeCommandResponse_interactive :: Lens' ExecuteCommandResponse (Maybe Bool)

Whether or not the execute command session is running in interactive mode. Amazon ECS only supports initiating interactive sessions, so you must specify true for this value.

executeCommandResponse_containerArn :: Lens' ExecuteCommandResponse (Maybe Text)

The Amazon Resource Name (ARN) of the container.

executeCommandResponse_session :: Lens' ExecuteCommandResponse (Maybe Session)

The details of the SSM session that was created for this instance of execute-command.

executeCommandResponse_containerName :: Lens' ExecuteCommandResponse (Maybe Text)

The name of the container.

executeCommandResponse_taskArn :: Lens' ExecuteCommandResponse (Maybe Text)

The Amazon Resource Name (ARN) of the task.

executeCommandResponse_httpStatus :: Lens' ExecuteCommandResponse Int

The response's http status code.