Network.AWS.ECS.ListAttributes

Description

Lists the attributes for Amazon ECS resources within a specified target type and cluster. When you specify a target type and cluster, ListAttributes returns a list of attribute objects, one for each attribute on each resource. You can filter the list of results to a single attribute name to only return results that have that name. You can also filter the results by attribute name and value, for example, to see which container instances in a cluster are running a Linux AMI (ecs.os-type=linux).

This operation returns paginated results.

Creating a Request

data ListAttributes

See: newListAttributes smart constructor.

Constructors

ListAttributes'

Fields

attributeValue :: Maybe Text

The value of the attribute with which to filter results. You must also specify an attribute name to use this parameter.

nextToken :: Maybe Text

The nextToken value returned from a ListAttributes request indicating that more results are available to fulfill the request and further calls will be needed. If maxResults was provided, it is possible the number of results to be fewer than maxResults.

This token should be treated as an opaque identifier that is only used to retrieve the next items in a list and not for other programmatic purposes.

maxResults :: Maybe Int

The maximum number of cluster results returned by ListAttributes in paginated output. When this parameter is used, ListAttributes only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListAttributes request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then ListAttributes returns up to 100 results and a nextToken value if applicable.

attributeName :: Maybe Text

The name of the attribute with which to filter the results.

cluster :: Maybe Text

The short name or full Amazon Resource Name (ARN) of the cluster to list attributes. If you do not specify a cluster, the default cluster is assumed.

targetType :: TargetType

The type of the target with which to list attributes.

newListAttributes

Arguments

:: TargetType
$sel:targetType:ListAttributes'

-> ListAttributes

Create a value of ListAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributeValue:ListAttributes', listAttributes_attributeValue - The value of the attribute with which to filter results. You must also specify an attribute name to use this parameter.

$sel:nextToken:ListAttributes', listAttributes_nextToken - The nextToken value returned from a ListAttributes request indicating that more results are available to fulfill the request and further calls will be needed. If maxResults was provided, it is possible the number of results to be fewer than maxResults.

This token should be treated as an opaque identifier that is only used to retrieve the next items in a list and not for other programmatic purposes.

$sel:maxResults:ListAttributes', listAttributes_maxResults - The maximum number of cluster results returned by ListAttributes in paginated output. When this parameter is used, ListAttributes only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListAttributes request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then ListAttributes returns up to 100 results and a nextToken value if applicable.

$sel:attributeName:ListAttributes', listAttributes_attributeName - The name of the attribute with which to filter the results.

$sel:cluster:ListAttributes', listAttributes_cluster - The short name or full Amazon Resource Name (ARN) of the cluster to list attributes. If you do not specify a cluster, the default cluster is assumed.

$sel:targetType:ListAttributes', listAttributes_targetType - The type of the target with which to list attributes.

Request Lenses

listAttributes_attributeValue :: Lens' ListAttributes (Maybe Text)

The value of the attribute with which to filter results. You must also specify an attribute name to use this parameter.

listAttributes_nextToken :: Lens' ListAttributes (Maybe Text)

The nextToken value returned from a ListAttributes request indicating that more results are available to fulfill the request and further calls will be needed. If maxResults was provided, it is possible the number of results to be fewer than maxResults.

This token should be treated as an opaque identifier that is only used to retrieve the next items in a list and not for other programmatic purposes.

listAttributes_maxResults :: Lens' ListAttributes (Maybe Int)

The maximum number of cluster results returned by ListAttributes in paginated output. When this parameter is used, ListAttributes only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListAttributes request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then ListAttributes returns up to 100 results and a nextToken value if applicable.

listAttributes_attributeName :: Lens' ListAttributes (Maybe Text)

The name of the attribute with which to filter the results.

listAttributes_cluster :: Lens' ListAttributes (Maybe Text)

The short name or full Amazon Resource Name (ARN) of the cluster to list attributes. If you do not specify a cluster, the default cluster is assumed.

listAttributes_targetType :: Lens' ListAttributes TargetType

The type of the target with which to list attributes.

Destructuring the Response

data ListAttributesResponse

See: newListAttributesResponse smart constructor.

Constructors

ListAttributesResponse'

Fields

nextToken :: Maybe Text

The nextToken value to include in a future ListAttributes request. When the results of a ListAttributes request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

attributes :: Maybe [Attribute]

A list of attribute objects that meet the criteria of the request.

httpStatus :: Int

The response's http status code.

newListAttributesResponse

Arguments

:: Int
$sel:httpStatus:ListAttributesResponse'

-> ListAttributesResponse

Create a value of ListAttributesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAttributes', listAttributesResponse_nextToken - The nextToken value to include in a future ListAttributes request. When the results of a ListAttributes request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

$sel:attributes:ListAttributesResponse', listAttributesResponse_attributes - A list of attribute objects that meet the criteria of the request.

$sel:httpStatus:ListAttributesResponse', listAttributesResponse_httpStatus - The response's http status code.

Response Lenses

listAttributesResponse_nextToken :: Lens' ListAttributesResponse (Maybe Text)

The nextToken value to include in a future ListAttributes request. When the results of a ListAttributes request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

listAttributesResponse_attributes :: Lens' ListAttributesResponse (Maybe [Attribute])

A list of attribute objects that meet the criteria of the request.

listAttributesResponse_httpStatus :: Lens' ListAttributesResponse Int

The response's http status code.