Network.AWS.ECS.ListContainerInstances

Description

Returns a list of container instances in a specified cluster. You can filter the results of a ListContainerInstances operation with cluster query language statements inside the filter parameter. For more information, see Cluster Query Language in the Amazon Elastic Container Service Developer Guide.

This operation returns paginated results.

Creating a Request

data ListContainerInstances

See: newListContainerInstances smart constructor.

Constructors

ListContainerInstances'

Fields

nextToken :: Maybe Text

The nextToken value returned from a ListContainerInstances request indicating that more results are available to fulfill the request and further calls will be needed. If maxResults was provided, it is possible the number of results to be fewer than maxResults.

This token should be treated as an opaque identifier that is only used to retrieve the next items in a list and not for other programmatic purposes.

status :: Maybe ContainerInstanceStatus

Filters the container instances by status. For example, if you specify the DRAINING status, the results include only container instances that have been set to DRAINING using UpdateContainerInstancesState. If you do not specify this parameter, the default is to include container instances set to all states other than INACTIVE.

maxResults :: Maybe Int

The maximum number of container instance results returned by ListContainerInstances in paginated output. When this parameter is used, ListContainerInstances only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListContainerInstances request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then ListContainerInstances returns up to 100 results and a nextToken value if applicable.

filter' :: Maybe Text

You can filter the results of a ListContainerInstances operation with cluster query language statements. For more information, see Cluster Query Language in the Amazon Elastic Container Service Developer Guide.

cluster :: Maybe Text

The short name or full Amazon Resource Name (ARN) of the cluster that hosts the container instances to list. If you do not specify a cluster, the default cluster is assumed.

newListContainerInstances :: ListContainerInstances

Create a value of ListContainerInstances with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListContainerInstances', listContainerInstances_nextToken - The nextToken value returned from a ListContainerInstances request indicating that more results are available to fulfill the request and further calls will be needed. If maxResults was provided, it is possible the number of results to be fewer than maxResults.

This token should be treated as an opaque identifier that is only used to retrieve the next items in a list and not for other programmatic purposes.

$sel:status:ListContainerInstances', listContainerInstances_status - Filters the container instances by status. For example, if you specify the DRAINING status, the results include only container instances that have been set to DRAINING using UpdateContainerInstancesState. If you do not specify this parameter, the default is to include container instances set to all states other than INACTIVE.

$sel:maxResults:ListContainerInstances', listContainerInstances_maxResults - The maximum number of container instance results returned by ListContainerInstances in paginated output. When this parameter is used, ListContainerInstances only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListContainerInstances request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then ListContainerInstances returns up to 100 results and a nextToken value if applicable.

$sel:filter':ListContainerInstances', listContainerInstances_filter - You can filter the results of a ListContainerInstances operation with cluster query language statements. For more information, see Cluster Query Language in the Amazon Elastic Container Service Developer Guide.

$sel:cluster:ListContainerInstances', listContainerInstances_cluster - The short name or full Amazon Resource Name (ARN) of the cluster that hosts the container instances to list. If you do not specify a cluster, the default cluster is assumed.

Request Lenses

listContainerInstances_nextToken :: Lens' ListContainerInstances (Maybe Text)

The nextToken value returned from a ListContainerInstances request indicating that more results are available to fulfill the request and further calls will be needed. If maxResults was provided, it is possible the number of results to be fewer than maxResults.

This token should be treated as an opaque identifier that is only used to retrieve the next items in a list and not for other programmatic purposes.

listContainerInstances_status :: Lens' ListContainerInstances (Maybe ContainerInstanceStatus)

Filters the container instances by status. For example, if you specify the DRAINING status, the results include only container instances that have been set to DRAINING using UpdateContainerInstancesState. If you do not specify this parameter, the default is to include container instances set to all states other than INACTIVE.

listContainerInstances_maxResults :: Lens' ListContainerInstances (Maybe Int)

The maximum number of container instance results returned by ListContainerInstances in paginated output. When this parameter is used, ListContainerInstances only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListContainerInstances request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then ListContainerInstances returns up to 100 results and a nextToken value if applicable.

listContainerInstances_filter :: Lens' ListContainerInstances (Maybe Text)

You can filter the results of a ListContainerInstances operation with cluster query language statements. For more information, see Cluster Query Language in the Amazon Elastic Container Service Developer Guide.

listContainerInstances_cluster :: Lens' ListContainerInstances (Maybe Text)

The short name or full Amazon Resource Name (ARN) of the cluster that hosts the container instances to list. If you do not specify a cluster, the default cluster is assumed.

Destructuring the Response

data ListContainerInstancesResponse

See: newListContainerInstancesResponse smart constructor.

Constructors

ListContainerInstancesResponse'

Fields

nextToken :: Maybe Text

The nextToken value to include in a future ListContainerInstances request. When the results of a ListContainerInstances request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

containerInstanceArns :: Maybe [Text]

The list of container instances with full ARN entries for each container instance associated with the specified cluster.

httpStatus :: Int

The response's http status code.

newListContainerInstancesResponse

Arguments

:: Int
$sel:httpStatus:ListContainerInstancesResponse'

-> ListContainerInstancesResponse

Create a value of ListContainerInstancesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListContainerInstances', listContainerInstancesResponse_nextToken - The nextToken value to include in a future ListContainerInstances request. When the results of a ListContainerInstances request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

$sel:containerInstanceArns:ListContainerInstancesResponse', listContainerInstancesResponse_containerInstanceArns - The list of container instances with full ARN entries for each container instance associated with the specified cluster.

$sel:httpStatus:ListContainerInstancesResponse', listContainerInstancesResponse_httpStatus - The response's http status code.

Response Lenses

listContainerInstancesResponse_nextToken :: Lens' ListContainerInstancesResponse (Maybe Text)

The nextToken value to include in a future ListContainerInstances request. When the results of a ListContainerInstances request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

listContainerInstancesResponse_containerInstanceArns :: Lens' ListContainerInstancesResponse (Maybe [Text])

The list of container instances with full ARN entries for each container instance associated with the specified cluster.

listContainerInstancesResponse_httpStatus :: Lens' ListContainerInstancesResponse Int

The response's http status code.