Network.AWS.ECS.ListServices

Description

Returns a list of services. You can filter the results by cluster, launch type, and scheduling strategy.

This operation returns paginated results.

Creating a Request

data ListServices

See: newListServices smart constructor.

Constructors

ListServices'

Fields

nextToken :: Maybe Text

The nextToken value returned from a ListServices request indicating that more results are available to fulfill the request and further calls will be needed. If maxResults was provided, it is possible the number of results to be fewer than maxResults.

This token should be treated as an opaque identifier that is only used to retrieve the next items in a list and not for other programmatic purposes.

maxResults :: Maybe Int

The maximum number of service results returned by ListServices in paginated output. When this parameter is used, ListServices only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListServices request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then ListServices returns up to 10 results and a nextToken value if applicable.

launchType :: Maybe LaunchType

The launch type to use when filtering the ListServices results.

schedulingStrategy :: Maybe SchedulingStrategy

The scheduling strategy to use when filtering the ListServices results.

cluster :: Maybe Text

The short name or full Amazon Resource Name (ARN) of the cluster to use when filtering the ListServices results. If you do not specify a cluster, the default cluster is assumed.

newListServices :: ListServices

Create a value of ListServices with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListServices', listServices_nextToken - The nextToken value returned from a ListServices request indicating that more results are available to fulfill the request and further calls will be needed. If maxResults was provided, it is possible the number of results to be fewer than maxResults.

This token should be treated as an opaque identifier that is only used to retrieve the next items in a list and not for other programmatic purposes.

$sel:maxResults:ListServices', listServices_maxResults - The maximum number of service results returned by ListServices in paginated output. When this parameter is used, ListServices only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListServices request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then ListServices returns up to 10 results and a nextToken value if applicable.

$sel:launchType:ListServices', listServices_launchType - The launch type to use when filtering the ListServices results.

$sel:schedulingStrategy:ListServices', listServices_schedulingStrategy - The scheduling strategy to use when filtering the ListServices results.

$sel:cluster:ListServices', listServices_cluster - The short name or full Amazon Resource Name (ARN) of the cluster to use when filtering the ListServices results. If you do not specify a cluster, the default cluster is assumed.

Request Lenses

listServices_nextToken :: Lens' ListServices (Maybe Text)

The nextToken value returned from a ListServices request indicating that more results are available to fulfill the request and further calls will be needed. If maxResults was provided, it is possible the number of results to be fewer than maxResults.

This token should be treated as an opaque identifier that is only used to retrieve the next items in a list and not for other programmatic purposes.

listServices_maxResults :: Lens' ListServices (Maybe Int)

The maximum number of service results returned by ListServices in paginated output. When this parameter is used, ListServices only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListServices request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then ListServices returns up to 10 results and a nextToken value if applicable.

listServices_launchType :: Lens' ListServices (Maybe LaunchType)

The launch type to use when filtering the ListServices results.

listServices_schedulingStrategy :: Lens' ListServices (Maybe SchedulingStrategy)

The scheduling strategy to use when filtering the ListServices results.

listServices_cluster :: Lens' ListServices (Maybe Text)

The short name or full Amazon Resource Name (ARN) of the cluster to use when filtering the ListServices results. If you do not specify a cluster, the default cluster is assumed.

Destructuring the Response

data ListServicesResponse

See: newListServicesResponse smart constructor.

Constructors

ListServicesResponse'

Fields

nextToken :: Maybe Text

The nextToken value to include in a future ListServices request. When the results of a ListServices request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

serviceArns :: Maybe [Text]

The list of full ARN entries for each service associated with the specified cluster.

httpStatus :: Int

The response's http status code.

newListServicesResponse

Arguments

:: Int
$sel:httpStatus:ListServicesResponse'

-> ListServicesResponse

Create a value of ListServicesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListServices', listServicesResponse_nextToken - The nextToken value to include in a future ListServices request. When the results of a ListServices request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

$sel:serviceArns:ListServicesResponse', listServicesResponse_serviceArns - The list of full ARN entries for each service associated with the specified cluster.

$sel:httpStatus:ListServicesResponse', listServicesResponse_httpStatus - The response's http status code.

Response Lenses

listServicesResponse_nextToken :: Lens' ListServicesResponse (Maybe Text)

The nextToken value to include in a future ListServices request. When the results of a ListServices request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

listServicesResponse_serviceArns :: Lens' ListServicesResponse (Maybe [Text])

The list of full ARN entries for each service associated with the specified cluster.

listServicesResponse_httpStatus :: Lens' ListServicesResponse Int

The response's http status code.