Network.AWS.ECS.ListTaskDefinitionFamilies

Description

Returns a list of task definition families that are registered to your account (which may include task definition families that no longer have any ACTIVE task definition revisions).

You can filter out task definition families that do not contain any ACTIVE task definition revisions by setting the status parameter to ACTIVE. You can also filter the results with the familyPrefix parameter.

This operation returns paginated results.

Creating a Request

data ListTaskDefinitionFamilies

See: newListTaskDefinitionFamilies smart constructor.

Constructors

ListTaskDefinitionFamilies'

Fields

status :: Maybe TaskDefinitionFamilyStatus

The task definition family status with which to filter the ListTaskDefinitionFamilies results. By default, both ACTIVE and INACTIVE task definition families are listed. If this parameter is set to ACTIVE, only task definition families that have an ACTIVE task definition revision are returned. If this parameter is set to INACTIVE, only task definition families that do not have any ACTIVE task definition revisions are returned. If you paginate the resulting output, be sure to keep the status value constant in each subsequent request.

familyPrefix :: Maybe Text

The familyPrefix is a string that is used to filter the results of ListTaskDefinitionFamilies. If you specify a familyPrefix, only task definition family names that begin with the familyPrefix string are returned.

nextToken :: Maybe Text

The nextToken value returned from a ListTaskDefinitionFamilies request indicating that more results are available to fulfill the request and further calls will be needed. If maxResults was provided, it is possible the number of results to be fewer than maxResults.

This token should be treated as an opaque identifier that is only used to retrieve the next items in a list and not for other programmatic purposes.

maxResults :: Maybe Int

The maximum number of task definition family results returned by ListTaskDefinitionFamilies in paginated output. When this parameter is used, ListTaskDefinitions only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListTaskDefinitionFamilies request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then ListTaskDefinitionFamilies returns up to 100 results and a nextToken value if applicable.

newListTaskDefinitionFamilies :: ListTaskDefinitionFamilies

Create a value of ListTaskDefinitionFamilies with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ListTaskDefinitionFamilies', listTaskDefinitionFamilies_status - The task definition family status with which to filter the ListTaskDefinitionFamilies results. By default, both ACTIVE and INACTIVE task definition families are listed. If this parameter is set to ACTIVE, only task definition families that have an ACTIVE task definition revision are returned. If this parameter is set to INACTIVE, only task definition families that do not have any ACTIVE task definition revisions are returned. If you paginate the resulting output, be sure to keep the status value constant in each subsequent request.

$sel:familyPrefix:ListTaskDefinitionFamilies', listTaskDefinitionFamilies_familyPrefix - The familyPrefix is a string that is used to filter the results of ListTaskDefinitionFamilies. If you specify a familyPrefix, only task definition family names that begin with the familyPrefix string are returned.

$sel:nextToken:ListTaskDefinitionFamilies', listTaskDefinitionFamilies_nextToken - The nextToken value returned from a ListTaskDefinitionFamilies request indicating that more results are available to fulfill the request and further calls will be needed. If maxResults was provided, it is possible the number of results to be fewer than maxResults.

This token should be treated as an opaque identifier that is only used to retrieve the next items in a list and not for other programmatic purposes.

$sel:maxResults:ListTaskDefinitionFamilies', listTaskDefinitionFamilies_maxResults - The maximum number of task definition family results returned by ListTaskDefinitionFamilies in paginated output. When this parameter is used, ListTaskDefinitions only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListTaskDefinitionFamilies request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then ListTaskDefinitionFamilies returns up to 100 results and a nextToken value if applicable.

Request Lenses

listTaskDefinitionFamilies_status :: Lens' ListTaskDefinitionFamilies (Maybe TaskDefinitionFamilyStatus)

The task definition family status with which to filter the ListTaskDefinitionFamilies results. By default, both ACTIVE and INACTIVE task definition families are listed. If this parameter is set to ACTIVE, only task definition families that have an ACTIVE task definition revision are returned. If this parameter is set to INACTIVE, only task definition families that do not have any ACTIVE task definition revisions are returned. If you paginate the resulting output, be sure to keep the status value constant in each subsequent request.

listTaskDefinitionFamilies_familyPrefix :: Lens' ListTaskDefinitionFamilies (Maybe Text)

The familyPrefix is a string that is used to filter the results of ListTaskDefinitionFamilies. If you specify a familyPrefix, only task definition family names that begin with the familyPrefix string are returned.

listTaskDefinitionFamilies_nextToken :: Lens' ListTaskDefinitionFamilies (Maybe Text)

The nextToken value returned from a ListTaskDefinitionFamilies request indicating that more results are available to fulfill the request and further calls will be needed. If maxResults was provided, it is possible the number of results to be fewer than maxResults.

This token should be treated as an opaque identifier that is only used to retrieve the next items in a list and not for other programmatic purposes.

listTaskDefinitionFamilies_maxResults :: Lens' ListTaskDefinitionFamilies (Maybe Int)

The maximum number of task definition family results returned by ListTaskDefinitionFamilies in paginated output. When this parameter is used, ListTaskDefinitions only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListTaskDefinitionFamilies request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then ListTaskDefinitionFamilies returns up to 100 results and a nextToken value if applicable.

Destructuring the Response

data ListTaskDefinitionFamiliesResponse

See: newListTaskDefinitionFamiliesResponse smart constructor.

Constructors

ListTaskDefinitionFamiliesResponse'

Fields

families :: Maybe [Text]

The list of task definition family names that match the ListTaskDefinitionFamilies request.

nextToken :: Maybe Text

The nextToken value to include in a future ListTaskDefinitionFamilies request. When the results of a ListTaskDefinitionFamilies request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

httpStatus :: Int

The response's http status code.

newListTaskDefinitionFamiliesResponse

Arguments

:: Int
$sel:httpStatus:ListTaskDefinitionFamiliesResponse'

-> ListTaskDefinitionFamiliesResponse

Create a value of ListTaskDefinitionFamiliesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:families:ListTaskDefinitionFamiliesResponse', listTaskDefinitionFamiliesResponse_families - The list of task definition family names that match the ListTaskDefinitionFamilies request.

$sel:nextToken:ListTaskDefinitionFamilies', listTaskDefinitionFamiliesResponse_nextToken - The nextToken value to include in a future ListTaskDefinitionFamilies request. When the results of a ListTaskDefinitionFamilies request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

$sel:httpStatus:ListTaskDefinitionFamiliesResponse', listTaskDefinitionFamiliesResponse_httpStatus - The response's http status code.

Response Lenses

listTaskDefinitionFamiliesResponse_families :: Lens' ListTaskDefinitionFamiliesResponse (Maybe [Text])

The list of task definition family names that match the ListTaskDefinitionFamilies request.

listTaskDefinitionFamiliesResponse_nextToken :: Lens' ListTaskDefinitionFamiliesResponse (Maybe Text)

The nextToken value to include in a future ListTaskDefinitionFamilies request. When the results of a ListTaskDefinitionFamilies request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

listTaskDefinitionFamiliesResponse_httpStatus :: Lens' ListTaskDefinitionFamiliesResponse Int

The response's http status code.