Network.AWS.ECS.ListTaskDefinitions

Description

Returns a list of task definitions that are registered to your account. You can filter the results by family name with the familyPrefix parameter or by status with the status parameter.

This operation returns paginated results.

Creating a Request

data ListTaskDefinitions

See: newListTaskDefinitions smart constructor.

Constructors

ListTaskDefinitions'

Fields

nextToken :: Maybe Text

The nextToken value returned from a ListTaskDefinitions request indicating that more results are available to fulfill the request and further calls will be needed. If maxResults was provided, it is possible the number of results to be fewer than maxResults.

This token should be treated as an opaque identifier that is only used to retrieve the next items in a list and not for other programmatic purposes.

status :: Maybe TaskDefinitionStatus

The task definition status with which to filter the ListTaskDefinitions results. By default, only ACTIVE task definitions are listed. By setting this parameter to INACTIVE, you can view task definitions that are INACTIVE as long as an active task or service still references them. If you paginate the resulting output, be sure to keep the status value constant in each subsequent request.

maxResults :: Maybe Int

The maximum number of task definition results returned by ListTaskDefinitions in paginated output. When this parameter is used, ListTaskDefinitions only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListTaskDefinitions request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then ListTaskDefinitions returns up to 100 results and a nextToken value if applicable.

familyPrefix :: Maybe Text

The full family name with which to filter the ListTaskDefinitions results. Specifying a familyPrefix limits the listed task definitions to task definition revisions that belong to that family.

sort :: Maybe SortOrder

The order in which to sort the results. Valid values are ASC and DESC. By default (ASC), task definitions are listed lexicographically by family name and in ascending numerical order by revision so that the newest task definitions in a family are listed last. Setting this parameter to DESC reverses the sort order on family name and revision so that the newest task definitions in a family are listed first.

newListTaskDefinitions :: ListTaskDefinitions

Create a value of ListTaskDefinitions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListTaskDefinitions', listTaskDefinitions_nextToken - The nextToken value returned from a ListTaskDefinitions request indicating that more results are available to fulfill the request and further calls will be needed. If maxResults was provided, it is possible the number of results to be fewer than maxResults.

This token should be treated as an opaque identifier that is only used to retrieve the next items in a list and not for other programmatic purposes.

$sel:status:ListTaskDefinitions', listTaskDefinitions_status - The task definition status with which to filter the ListTaskDefinitions results. By default, only ACTIVE task definitions are listed. By setting this parameter to INACTIVE, you can view task definitions that are INACTIVE as long as an active task or service still references them. If you paginate the resulting output, be sure to keep the status value constant in each subsequent request.

$sel:maxResults:ListTaskDefinitions', listTaskDefinitions_maxResults - The maximum number of task definition results returned by ListTaskDefinitions in paginated output. When this parameter is used, ListTaskDefinitions only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListTaskDefinitions request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then ListTaskDefinitions returns up to 100 results and a nextToken value if applicable.

$sel:familyPrefix:ListTaskDefinitions', listTaskDefinitions_familyPrefix - The full family name with which to filter the ListTaskDefinitions results. Specifying a familyPrefix limits the listed task definitions to task definition revisions that belong to that family.

$sel:sort:ListTaskDefinitions', listTaskDefinitions_sort - The order in which to sort the results. Valid values are ASC and DESC. By default (ASC), task definitions are listed lexicographically by family name and in ascending numerical order by revision so that the newest task definitions in a family are listed last. Setting this parameter to DESC reverses the sort order on family name and revision so that the newest task definitions in a family are listed first.

Request Lenses

listTaskDefinitions_nextToken :: Lens' ListTaskDefinitions (Maybe Text)

The nextToken value returned from a ListTaskDefinitions request indicating that more results are available to fulfill the request and further calls will be needed. If maxResults was provided, it is possible the number of results to be fewer than maxResults.

This token should be treated as an opaque identifier that is only used to retrieve the next items in a list and not for other programmatic purposes.

listTaskDefinitions_status :: Lens' ListTaskDefinitions (Maybe TaskDefinitionStatus)

The task definition status with which to filter the ListTaskDefinitions results. By default, only ACTIVE task definitions are listed. By setting this parameter to INACTIVE, you can view task definitions that are INACTIVE as long as an active task or service still references them. If you paginate the resulting output, be sure to keep the status value constant in each subsequent request.

listTaskDefinitions_maxResults :: Lens' ListTaskDefinitions (Maybe Int)

The maximum number of task definition results returned by ListTaskDefinitions in paginated output. When this parameter is used, ListTaskDefinitions only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListTaskDefinitions request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then ListTaskDefinitions returns up to 100 results and a nextToken value if applicable.

listTaskDefinitions_familyPrefix :: Lens' ListTaskDefinitions (Maybe Text)

The full family name with which to filter the ListTaskDefinitions results. Specifying a familyPrefix limits the listed task definitions to task definition revisions that belong to that family.

listTaskDefinitions_sort :: Lens' ListTaskDefinitions (Maybe SortOrder)

The order in which to sort the results. Valid values are ASC and DESC. By default (ASC), task definitions are listed lexicographically by family name and in ascending numerical order by revision so that the newest task definitions in a family are listed last. Setting this parameter to DESC reverses the sort order on family name and revision so that the newest task definitions in a family are listed first.

Destructuring the Response

data ListTaskDefinitionsResponse

See: newListTaskDefinitionsResponse smart constructor.

Constructors

ListTaskDefinitionsResponse'

Fields

nextToken :: Maybe Text

The nextToken value to include in a future ListTaskDefinitions request. When the results of a ListTaskDefinitions request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

taskDefinitionArns :: Maybe [Text]

The list of task definition Amazon Resource Name (ARN) entries for the ListTaskDefinitions request.

httpStatus :: Int

The response's http status code.

newListTaskDefinitionsResponse

Arguments

:: Int
$sel:httpStatus:ListTaskDefinitionsResponse'

-> ListTaskDefinitionsResponse

Create a value of ListTaskDefinitionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListTaskDefinitions', listTaskDefinitionsResponse_nextToken - The nextToken value to include in a future ListTaskDefinitions request. When the results of a ListTaskDefinitions request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

$sel:taskDefinitionArns:ListTaskDefinitionsResponse', listTaskDefinitionsResponse_taskDefinitionArns - The list of task definition Amazon Resource Name (ARN) entries for the ListTaskDefinitions request.

$sel:httpStatus:ListTaskDefinitionsResponse', listTaskDefinitionsResponse_httpStatus - The response's http status code.

Response Lenses

listTaskDefinitionsResponse_nextToken :: Lens' ListTaskDefinitionsResponse (Maybe Text)

The nextToken value to include in a future ListTaskDefinitions request. When the results of a ListTaskDefinitions request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

listTaskDefinitionsResponse_taskDefinitionArns :: Lens' ListTaskDefinitionsResponse (Maybe [Text])

The list of task definition Amazon Resource Name (ARN) entries for the ListTaskDefinitions request.

listTaskDefinitionsResponse_httpStatus :: Lens' ListTaskDefinitionsResponse Int

The response's http status code.