Network.AWS.ECS.PutAccountSetting

Description

Modifies an account setting. Account settings are set on a per-Region basis.

If you change the account setting for the root user, the default settings for all of the IAM users and roles for which no individual account setting has been specified are reset. For more information, see Account Settings in the Amazon Elastic Container Service Developer Guide.

When serviceLongArnFormat, taskLongArnFormat, or containerInstanceLongArnFormat are specified, the Amazon Resource Name (ARN) and resource ID format of the resource type for a specified IAM user, IAM role, or the root user for an account is affected. The opt-in and opt-out account setting must be set for each Amazon ECS resource separately. The ARN and resource ID format of a resource will be defined by the opt-in status of the IAM user or role that created the resource. You must enable this setting to use Amazon ECS features such as resource tagging.

When awsvpcTrunking is specified, the elastic network interface (ENI) limit for any new container instances that support the feature is changed. If awsvpcTrunking is enabled, any new container instances that support the feature are launched have the increased ENI limits available to them. For more information, see Elastic Network Interface Trunking in the Amazon Elastic Container Service Developer Guide.

When containerInsights is specified, the default setting indicating whether CloudWatch Container Insights is enabled for your clusters is changed. If containerInsights is enabled, any new clusters that are created will have Container Insights enabled unless you disable it during cluster creation. For more information, see CloudWatch Container Insights in the Amazon Elastic Container Service Developer Guide.

Creating a Request

data PutAccountSetting

See: newPutAccountSetting smart constructor.

Constructors

PutAccountSetting'

Fields

principalArn :: Maybe Text

The ARN of the principal, which can be an IAM user, IAM role, or the root user. If you specify the root user, it modifies the account setting for all IAM users, IAM roles, and the root user of the account unless an IAM user or role explicitly overrides these settings. If this field is omitted, the setting is changed only for the authenticated user.

Federated users assume the account setting of the root user and can't have explicit account settings set for them.

name :: SettingName

The Amazon ECS resource name for which to modify the account setting. If serviceLongArnFormat is specified, the ARN for your Amazon ECS services is affected. If taskLongArnFormat is specified, the ARN and resource ID for your Amazon ECS tasks is affected. If containerInstanceLongArnFormat is specified, the ARN and resource ID for your Amazon ECS container instances is affected. If awsvpcTrunking is specified, the elastic network interface (ENI) limit for your Amazon ECS container instances is affected. If containerInsights is specified, the default setting for CloudWatch Container Insights for your clusters is affected.

value :: Text

The account setting value for the specified principal ARN. Accepted values are enabled and disabled.

newPutAccountSetting

Arguments

:: SettingName
$sel:name:PutAccountSetting'

-> Text
$sel:value:PutAccountSetting'

-> PutAccountSetting

Create a value of PutAccountSetting with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:principalArn:PutAccountSetting', putAccountSetting_principalArn - The ARN of the principal, which can be an IAM user, IAM role, or the root user. If you specify the root user, it modifies the account setting for all IAM users, IAM roles, and the root user of the account unless an IAM user or role explicitly overrides these settings. If this field is omitted, the setting is changed only for the authenticated user.

Federated users assume the account setting of the root user and can't have explicit account settings set for them.

$sel:name:PutAccountSetting', putAccountSetting_name - The Amazon ECS resource name for which to modify the account setting. If serviceLongArnFormat is specified, the ARN for your Amazon ECS services is affected. If taskLongArnFormat is specified, the ARN and resource ID for your Amazon ECS tasks is affected. If containerInstanceLongArnFormat is specified, the ARN and resource ID for your Amazon ECS container instances is affected. If awsvpcTrunking is specified, the elastic network interface (ENI) limit for your Amazon ECS container instances is affected. If containerInsights is specified, the default setting for CloudWatch Container Insights for your clusters is affected.

$sel:value:PutAccountSetting', putAccountSetting_value - The account setting value for the specified principal ARN. Accepted values are enabled and disabled.

Request Lenses

putAccountSetting_principalArn :: Lens' PutAccountSetting (Maybe Text)

The ARN of the principal, which can be an IAM user, IAM role, or the root user. If you specify the root user, it modifies the account setting for all IAM users, IAM roles, and the root user of the account unless an IAM user or role explicitly overrides these settings. If this field is omitted, the setting is changed only for the authenticated user.

Federated users assume the account setting of the root user and can't have explicit account settings set for them.

putAccountSetting_name :: Lens' PutAccountSetting SettingName

The Amazon ECS resource name for which to modify the account setting. If serviceLongArnFormat is specified, the ARN for your Amazon ECS services is affected. If taskLongArnFormat is specified, the ARN and resource ID for your Amazon ECS tasks is affected. If containerInstanceLongArnFormat is specified, the ARN and resource ID for your Amazon ECS container instances is affected. If awsvpcTrunking is specified, the elastic network interface (ENI) limit for your Amazon ECS container instances is affected. If containerInsights is specified, the default setting for CloudWatch Container Insights for your clusters is affected.

putAccountSetting_value :: Lens' PutAccountSetting Text

The account setting value for the specified principal ARN. Accepted values are enabled and disabled.

Destructuring the Response

data PutAccountSettingResponse

See: newPutAccountSettingResponse smart constructor.

Constructors

PutAccountSettingResponse'

Fields

setting :: Maybe Setting

The current account setting for a resource.

httpStatus :: Int

The response's http status code.

newPutAccountSettingResponse

Arguments

:: Int
$sel:httpStatus:PutAccountSettingResponse'

-> PutAccountSettingResponse

Create a value of PutAccountSettingResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:setting:PutAccountSettingResponse', putAccountSettingResponse_setting - The current account setting for a resource.

$sel:httpStatus:PutAccountSettingResponse', putAccountSettingResponse_httpStatus - The response's http status code.

Response Lenses

putAccountSettingResponse_setting :: Lens' PutAccountSettingResponse (Maybe Setting)

The current account setting for a resource.

putAccountSettingResponse_httpStatus :: Lens' PutAccountSettingResponse Int

The response's http status code.