Network.AWS.ECS.RegisterContainerInstance

Description

This action is only used by the Amazon ECS agent, and it is not intended for use outside of the agent.

Registers an EC2 instance into the specified cluster. This instance becomes available to place containers on.

Creating a Request

data RegisterContainerInstance

See: newRegisterContainerInstance smart constructor.

Constructors

RegisterContainerInstance'

Fields

platformDevices :: Maybe [PlatformDevice]

The devices that are available on the container instance. The only supported device type is a GPU.

instanceIdentityDocumentSignature :: Maybe Text

The instance identity document signature for the EC2 instance to register. This signature can be found by running the following command from the instance: curl http://169.254.169.254/latest/dynamic/instance-identity/signature/

cluster :: Maybe Text

The short name or full Amazon Resource Name (ARN) of the cluster with which to register your container instance. If you do not specify a cluster, the default cluster is assumed.

instanceIdentityDocument :: Maybe Text

The instance identity document for the EC2 instance to register. This document can be found by running the following command from the instance: curl http://169.254.169.254/latest/dynamic/instance-identity/document/

containerInstanceArn :: Maybe Text

The ARN of the container instance (if it was previously registered).

versionInfo :: Maybe VersionInfo

The version information for the Amazon ECS container agent and Docker daemon running on the container instance.

attributes :: Maybe [Attribute]

The container instance attributes that this container instance supports.

totalResources :: Maybe [Resource]

The resources available on the instance.

tags :: Maybe [Tag]

The metadata that you apply to the container instance to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define.

The following basic restrictions apply to tags:

Maximum number of tags per resource - 50

For each resource, each tag key must be unique, and each tag key can have only one value.

Maximum key length - 128 Unicode characters in UTF-8

Maximum value length - 256 Unicode characters in UTF-8

If your tagging schema is used across multiple services and resources, remember that other services may have restrictions on allowed characters. Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following characters: + - = . _ : / @.

Tag keys and values are case-sensitive.

Do not use aws:, AWS:, or any upper or lowercase combination of such as a prefix for either keys or values as it is reserved for Amazon Web Services use. You cannot edit or delete tag keys or values with this prefix. Tags with this prefix do not count against your tags per resource limit.

newRegisterContainerInstance :: RegisterContainerInstance

Create a value of RegisterContainerInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:platformDevices:RegisterContainerInstance', registerContainerInstance_platformDevices - The devices that are available on the container instance. The only supported device type is a GPU.

$sel:instanceIdentityDocumentSignature:RegisterContainerInstance', registerContainerInstance_instanceIdentityDocumentSignature - The instance identity document signature for the EC2 instance to register. This signature can be found by running the following command from the instance: curl http://169.254.169.254/latest/dynamic/instance-identity/signature/

$sel:cluster:RegisterContainerInstance', registerContainerInstance_cluster - The short name or full Amazon Resource Name (ARN) of the cluster with which to register your container instance. If you do not specify a cluster, the default cluster is assumed.

$sel:instanceIdentityDocument:RegisterContainerInstance', registerContainerInstance_instanceIdentityDocument - The instance identity document for the EC2 instance to register. This document can be found by running the following command from the instance: curl http://169.254.169.254/latest/dynamic/instance-identity/document/

$sel:containerInstanceArn:RegisterContainerInstance', registerContainerInstance_containerInstanceArn - The ARN of the container instance (if it was previously registered).

$sel:versionInfo:RegisterContainerInstance', registerContainerInstance_versionInfo - The version information for the Amazon ECS container agent and Docker daemon running on the container instance.

$sel:attributes:RegisterContainerInstance', registerContainerInstance_attributes - The container instance attributes that this container instance supports.

$sel:totalResources:RegisterContainerInstance', registerContainerInstance_totalResources - The resources available on the instance.

$sel:tags:RegisterContainerInstance', registerContainerInstance_tags - The metadata that you apply to the container instance to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define.

The following basic restrictions apply to tags:

Maximum number of tags per resource - 50

For each resource, each tag key must be unique, and each tag key can have only one value.

Maximum key length - 128 Unicode characters in UTF-8

Maximum value length - 256 Unicode characters in UTF-8

If your tagging schema is used across multiple services and resources, remember that other services may have restrictions on allowed characters. Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following characters: + - = . _ : / @.

Tag keys and values are case-sensitive.

Do not use aws:, AWS:, or any upper or lowercase combination of such as a prefix for either keys or values as it is reserved for Amazon Web Services use. You cannot edit or delete tag keys or values with this prefix. Tags with this prefix do not count against your tags per resource limit.

Request Lenses

registerContainerInstance_platformDevices :: Lens' RegisterContainerInstance (Maybe [PlatformDevice])

The devices that are available on the container instance. The only supported device type is a GPU.

registerContainerInstance_instanceIdentityDocumentSignature :: Lens' RegisterContainerInstance (Maybe Text)

The instance identity document signature for the EC2 instance to register. This signature can be found by running the following command from the instance: curl http://169.254.169.254/latest/dynamic/instance-identity/signature/

registerContainerInstance_cluster :: Lens' RegisterContainerInstance (Maybe Text)

The short name or full Amazon Resource Name (ARN) of the cluster with which to register your container instance. If you do not specify a cluster, the default cluster is assumed.

registerContainerInstance_instanceIdentityDocument :: Lens' RegisterContainerInstance (Maybe Text)

The instance identity document for the EC2 instance to register. This document can be found by running the following command from the instance: curl http://169.254.169.254/latest/dynamic/instance-identity/document/

registerContainerInstance_containerInstanceArn :: Lens' RegisterContainerInstance (Maybe Text)

The ARN of the container instance (if it was previously registered).

registerContainerInstance_versionInfo :: Lens' RegisterContainerInstance (Maybe VersionInfo)

The version information for the Amazon ECS container agent and Docker daemon running on the container instance.

registerContainerInstance_attributes :: Lens' RegisterContainerInstance (Maybe [Attribute])

The container instance attributes that this container instance supports.

registerContainerInstance_totalResources :: Lens' RegisterContainerInstance (Maybe [Resource])

The resources available on the instance.

registerContainerInstance_tags :: Lens' RegisterContainerInstance (Maybe [Tag])

The metadata that you apply to the container instance to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define.

The following basic restrictions apply to tags:

Maximum number of tags per resource - 50

For each resource, each tag key must be unique, and each tag key can have only one value.

Maximum key length - 128 Unicode characters in UTF-8

Maximum value length - 256 Unicode characters in UTF-8

If your tagging schema is used across multiple services and resources, remember that other services may have restrictions on allowed characters. Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following characters: + - = . _ : / @.

Tag keys and values are case-sensitive.

Do not use aws:, AWS:, or any upper or lowercase combination of such as a prefix for either keys or values as it is reserved for Amazon Web Services use. You cannot edit or delete tag keys or values with this prefix. Tags with this prefix do not count against your tags per resource limit.

Destructuring the Response

data RegisterContainerInstanceResponse

See: newRegisterContainerInstanceResponse smart constructor.

Constructors

RegisterContainerInstanceResponse'

Fields

containerInstance :: Maybe ContainerInstance

The container instance that was registered.

httpStatus :: Int

The response's http status code.

newRegisterContainerInstanceResponse

Arguments

:: Int
$sel:httpStatus:RegisterContainerInstanceResponse'

-> RegisterContainerInstanceResponse

Create a value of RegisterContainerInstanceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:containerInstance:RegisterContainerInstanceResponse', registerContainerInstanceResponse_containerInstance - The container instance that was registered.

$sel:httpStatus:RegisterContainerInstanceResponse', registerContainerInstanceResponse_httpStatus - The response's http status code.

Response Lenses

registerContainerInstanceResponse_containerInstance :: Lens' RegisterContainerInstanceResponse (Maybe ContainerInstance)

The container instance that was registered.

registerContainerInstanceResponse_httpStatus :: Lens' RegisterContainerInstanceResponse Int

The response's http status code.