Network.AWS.ECS.StartTask

Description

Starts a new task from the specified task definition on the specified container instance or instances.

Alternatively, you can use RunTask to place tasks for you. For more information, see Scheduling Tasks in the Amazon Elastic Container Service Developer Guide.

Creating a Request

data StartTask

See: newStartTask smart constructor.

Constructors

StartTask'

Fields

referenceId :: Maybe Text

The reference ID to use for the task.

networkConfiguration :: Maybe NetworkConfiguration

The VPC subnet and security group configuration for tasks that receive their own elastic network interface by using the awsvpc networking mode.

enableECSManagedTags :: Maybe Bool

Specifies whether to enable Amazon ECS managed tags for the task. For more information, see Tagging Your Amazon ECS Resources in the Amazon Elastic Container Service Developer Guide.

startedBy :: Maybe Text

An optional tag specified when a task is started. For example, if you automatically trigger a task to run a batch process job, you could apply a unique identifier for that job to your task with the startedBy parameter. You can then identify which tasks belong to that job by filtering the results of a ListTasks call with the startedBy value. Up to 36 letters (uppercase and lowercase), numbers, hyphens, and underscores are allowed.

If a task is started by an Amazon ECS service, then the startedBy parameter contains the deployment ID of the service that starts it.

group' :: Maybe Text

The name of the task group to associate with the task. The default value is the family name of the task definition (for example, family:my-family-name).

overrides :: Maybe TaskOverride

A list of container overrides in JSON format that specify the name of a container in the specified task definition and the overrides it should receive. You can override the default command for a container (that is specified in the task definition or Docker image) with a command override. You can also override existing environment variables (that are specified in the task definition or Docker image) on a container or add new environment variables to it with an environment override.

A total of 8192 characters are allowed for overrides. This limit includes the JSON formatting characters of the override structure.

enableExecuteCommand :: Maybe Bool

Whether or not the execute command functionality is enabled for the task. If true, this enables execute command functionality on all containers in the task.

tags :: Maybe [Tag]

The metadata that you apply to the task to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define.

The following basic restrictions apply to tags:

Maximum number of tags per resource - 50

For each resource, each tag key must be unique, and each tag key can have only one value.

Maximum key length - 128 Unicode characters in UTF-8

Maximum value length - 256 Unicode characters in UTF-8

If your tagging schema is used across multiple services and resources, remember that other services may have restrictions on allowed characters. Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following characters: + - = . _ : / @.

Tag keys and values are case-sensitive.

Do not use aws:, AWS:, or any upper or lowercase combination of such as a prefix for either keys or values as it is reserved for Amazon Web Services use. You cannot edit or delete tag keys or values with this prefix. Tags with this prefix do not count against your tags per resource limit.

cluster :: Maybe Text

The short name or full Amazon Resource Name (ARN) of the cluster on which to start your task. If you do not specify a cluster, the default cluster is assumed.

propagateTags :: Maybe PropagateTags

Specifies whether to propagate the tags from the task definition or the service to the task. If no value is specified, the tags are not propagated.

containerInstances :: [Text]

The container instance IDs or full ARN entries for the container instances on which you would like to place your task. You can specify up to 10 container instances.

taskDefinition :: Text

The family and revision (family:revision) or full ARN of the task definition to start. If a revision is not specified, the latest ACTIVE revision is used.

newStartTask

Arguments

:: Text
$sel:taskDefinition:StartTask'

-> StartTask

Create a value of StartTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:referenceId:StartTask', startTask_referenceId - The reference ID to use for the task.

$sel:networkConfiguration:StartTask', startTask_networkConfiguration - The VPC subnet and security group configuration for tasks that receive their own elastic network interface by using the awsvpc networking mode.

$sel:enableECSManagedTags:StartTask', startTask_enableECSManagedTags - Specifies whether to enable Amazon ECS managed tags for the task. For more information, see Tagging Your Amazon ECS Resources in the Amazon Elastic Container Service Developer Guide.

$sel:startedBy:StartTask', startTask_startedBy - An optional tag specified when a task is started. For example, if you automatically trigger a task to run a batch process job, you could apply a unique identifier for that job to your task with the startedBy parameter. You can then identify which tasks belong to that job by filtering the results of a ListTasks call with the startedBy value. Up to 36 letters (uppercase and lowercase), numbers, hyphens, and underscores are allowed.

If a task is started by an Amazon ECS service, then the startedBy parameter contains the deployment ID of the service that starts it.

$sel:group':StartTask', startTask_group - The name of the task group to associate with the task. The default value is the family name of the task definition (for example, family:my-family-name).

$sel:overrides:StartTask', startTask_overrides - A list of container overrides in JSON format that specify the name of a container in the specified task definition and the overrides it should receive. You can override the default command for a container (that is specified in the task definition or Docker image) with a command override. You can also override existing environment variables (that are specified in the task definition or Docker image) on a container or add new environment variables to it with an environment override.

A total of 8192 characters are allowed for overrides. This limit includes the JSON formatting characters of the override structure.

$sel:enableExecuteCommand:StartTask', startTask_enableExecuteCommand - Whether or not the execute command functionality is enabled for the task. If true, this enables execute command functionality on all containers in the task.

$sel:tags:StartTask', startTask_tags - The metadata that you apply to the task to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define.

The following basic restrictions apply to tags:

Maximum number of tags per resource - 50

For each resource, each tag key must be unique, and each tag key can have only one value.

Maximum key length - 128 Unicode characters in UTF-8

Maximum value length - 256 Unicode characters in UTF-8

If your tagging schema is used across multiple services and resources, remember that other services may have restrictions on allowed characters. Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following characters: + - = . _ : / @.

Tag keys and values are case-sensitive.

Do not use aws:, AWS:, or any upper or lowercase combination of such as a prefix for either keys or values as it is reserved for Amazon Web Services use. You cannot edit or delete tag keys or values with this prefix. Tags with this prefix do not count against your tags per resource limit.

$sel:cluster:StartTask', startTask_cluster - The short name or full Amazon Resource Name (ARN) of the cluster on which to start your task. If you do not specify a cluster, the default cluster is assumed.

$sel:propagateTags:StartTask', startTask_propagateTags - Specifies whether to propagate the tags from the task definition or the service to the task. If no value is specified, the tags are not propagated.

$sel:containerInstances:StartTask', startTask_containerInstances - The container instance IDs or full ARN entries for the container instances on which you would like to place your task. You can specify up to 10 container instances.

$sel:taskDefinition:StartTask', startTask_taskDefinition - The family and revision (family:revision) or full ARN of the task definition to start. If a revision is not specified, the latest ACTIVE revision is used.

Request Lenses

startTask_referenceId :: Lens' StartTask (Maybe Text)

The reference ID to use for the task.

startTask_networkConfiguration :: Lens' StartTask (Maybe NetworkConfiguration)

The VPC subnet and security group configuration for tasks that receive their own elastic network interface by using the awsvpc networking mode.

startTask_enableECSManagedTags :: Lens' StartTask (Maybe Bool)

Specifies whether to enable Amazon ECS managed tags for the task. For more information, see Tagging Your Amazon ECS Resources in the Amazon Elastic Container Service Developer Guide.

startTask_startedBy :: Lens' StartTask (Maybe Text)

An optional tag specified when a task is started. For example, if you automatically trigger a task to run a batch process job, you could apply a unique identifier for that job to your task with the startedBy parameter. You can then identify which tasks belong to that job by filtering the results of a ListTasks call with the startedBy value. Up to 36 letters (uppercase and lowercase), numbers, hyphens, and underscores are allowed.

If a task is started by an Amazon ECS service, then the startedBy parameter contains the deployment ID of the service that starts it.

startTask_group :: Lens' StartTask (Maybe Text)

The name of the task group to associate with the task. The default value is the family name of the task definition (for example, family:my-family-name).

startTask_overrides :: Lens' StartTask (Maybe TaskOverride)

A list of container overrides in JSON format that specify the name of a container in the specified task definition and the overrides it should receive. You can override the default command for a container (that is specified in the task definition or Docker image) with a command override. You can also override existing environment variables (that are specified in the task definition or Docker image) on a container or add new environment variables to it with an environment override.

A total of 8192 characters are allowed for overrides. This limit includes the JSON formatting characters of the override structure.

startTask_enableExecuteCommand :: Lens' StartTask (Maybe Bool)

Whether or not the execute command functionality is enabled for the task. If true, this enables execute command functionality on all containers in the task.

startTask_tags :: Lens' StartTask (Maybe [Tag])

The metadata that you apply to the task to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define.

The following basic restrictions apply to tags:

Maximum number of tags per resource - 50

For each resource, each tag key must be unique, and each tag key can have only one value.

Maximum key length - 128 Unicode characters in UTF-8

Maximum value length - 256 Unicode characters in UTF-8

If your tagging schema is used across multiple services and resources, remember that other services may have restrictions on allowed characters. Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following characters: + - = . _ : / @.

Tag keys and values are case-sensitive.

Do not use aws:, AWS:, or any upper or lowercase combination of such as a prefix for either keys or values as it is reserved for Amazon Web Services use. You cannot edit or delete tag keys or values with this prefix. Tags with this prefix do not count against your tags per resource limit.

startTask_cluster :: Lens' StartTask (Maybe Text)

The short name or full Amazon Resource Name (ARN) of the cluster on which to start your task. If you do not specify a cluster, the default cluster is assumed.

startTask_propagateTags :: Lens' StartTask (Maybe PropagateTags)

Specifies whether to propagate the tags from the task definition or the service to the task. If no value is specified, the tags are not propagated.

startTask_containerInstances :: Lens' StartTask [Text]

The container instance IDs or full ARN entries for the container instances on which you would like to place your task. You can specify up to 10 container instances.

startTask_taskDefinition :: Lens' StartTask Text

The family and revision (family:revision) or full ARN of the task definition to start. If a revision is not specified, the latest ACTIVE revision is used.

Destructuring the Response

data StartTaskResponse

See: newStartTaskResponse smart constructor.

Constructors

StartTaskResponse'

Fields

tasks :: Maybe [Task]

A full description of the tasks that were started. Each task that was successfully placed on your container instances is described.

failures :: Maybe [Failure]

Any failures associated with the call.

httpStatus :: Int

The response's http status code.

newStartTaskResponse

Arguments

:: Int
$sel:httpStatus:StartTaskResponse'

-> StartTaskResponse

Create a value of StartTaskResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tasks:StartTaskResponse', startTaskResponse_tasks - A full description of the tasks that were started. Each task that was successfully placed on your container instances is described.

$sel:failures:StartTaskResponse', startTaskResponse_failures - Any failures associated with the call.

$sel:httpStatus:StartTaskResponse', startTaskResponse_httpStatus - The response's http status code.

Response Lenses

startTaskResponse_tasks :: Lens' StartTaskResponse (Maybe [Task])

A full description of the tasks that were started. Each task that was successfully placed on your container instances is described.

startTaskResponse_failures :: Lens' StartTaskResponse (Maybe [Failure])

Any failures associated with the call.

startTaskResponse_httpStatus :: Lens' StartTaskResponse Int

The response's http status code.