Network.AWS.ECS.SubmitContainerStateChange

Description

This action is only used by the Amazon ECS agent, and it is not intended for use outside of the agent.

Sent to acknowledge that a container changed states.

Creating a Request

data SubmitContainerStateChange

See: newSubmitContainerStateChange smart constructor.

Constructors

SubmitContainerStateChange'

Fields

networkBindings :: Maybe [NetworkBinding]

The network bindings of the container.

status :: Maybe Text

The status of the state change request.

cluster :: Maybe Text

The short name or full ARN of the cluster that hosts the container.

containerName :: Maybe Text

The name of the container.

reason :: Maybe Text

The reason for the state change request.

exitCode :: Maybe Int

The exit code returned for the state change request.

task :: Maybe Text

The task ID or full Amazon Resource Name (ARN) of the task that hosts the container.

runtimeId :: Maybe Text

The ID of the Docker container.

newSubmitContainerStateChange :: SubmitContainerStateChange

Create a value of SubmitContainerStateChange with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:networkBindings:SubmitContainerStateChange', submitContainerStateChange_networkBindings - The network bindings of the container.

$sel:status:SubmitContainerStateChange', submitContainerStateChange_status - The status of the state change request.

$sel:cluster:SubmitContainerStateChange', submitContainerStateChange_cluster - The short name or full ARN of the cluster that hosts the container.

$sel:containerName:SubmitContainerStateChange', submitContainerStateChange_containerName - The name of the container.

$sel:reason:SubmitContainerStateChange', submitContainerStateChange_reason - The reason for the state change request.

$sel:exitCode:SubmitContainerStateChange', submitContainerStateChange_exitCode - The exit code returned for the state change request.

$sel:task:SubmitContainerStateChange', submitContainerStateChange_task - The task ID or full Amazon Resource Name (ARN) of the task that hosts the container.

$sel:runtimeId:SubmitContainerStateChange', submitContainerStateChange_runtimeId - The ID of the Docker container.

Request Lenses

submitContainerStateChange_networkBindings :: Lens' SubmitContainerStateChange (Maybe [NetworkBinding])

The network bindings of the container.

submitContainerStateChange_status :: Lens' SubmitContainerStateChange (Maybe Text)

The status of the state change request.

submitContainerStateChange_cluster :: Lens' SubmitContainerStateChange (Maybe Text)

The short name or full ARN of the cluster that hosts the container.

submitContainerStateChange_containerName :: Lens' SubmitContainerStateChange (Maybe Text)

The name of the container.

submitContainerStateChange_reason :: Lens' SubmitContainerStateChange (Maybe Text)

The reason for the state change request.

submitContainerStateChange_exitCode :: Lens' SubmitContainerStateChange (Maybe Int)

The exit code returned for the state change request.

submitContainerStateChange_task :: Lens' SubmitContainerStateChange (Maybe Text)

The task ID or full Amazon Resource Name (ARN) of the task that hosts the container.

submitContainerStateChange_runtimeId :: Lens' SubmitContainerStateChange (Maybe Text)

The ID of the Docker container.

Destructuring the Response

data SubmitContainerStateChangeResponse

See: newSubmitContainerStateChangeResponse smart constructor.

Constructors

SubmitContainerStateChangeResponse'

Fields

acknowledgment :: Maybe Text

Acknowledgement of the state change.

httpStatus :: Int

The response's http status code.

newSubmitContainerStateChangeResponse

Arguments

:: Int
$sel:httpStatus:SubmitContainerStateChangeResponse'

-> SubmitContainerStateChangeResponse

Create a value of SubmitContainerStateChangeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:acknowledgment:SubmitContainerStateChangeResponse', submitContainerStateChangeResponse_acknowledgment - Acknowledgement of the state change.

$sel:httpStatus:SubmitContainerStateChangeResponse', submitContainerStateChangeResponse_httpStatus - The response's http status code.

Response Lenses

submitContainerStateChangeResponse_acknowledgment :: Lens' SubmitContainerStateChangeResponse (Maybe Text)

Acknowledgement of the state change.

submitContainerStateChangeResponse_httpStatus :: Lens' SubmitContainerStateChangeResponse Int

The response's http status code.