Network.AWS.ECS.Types.Attribute

Description

Documentation

data Attribute

An attribute is a name-value pair associated with an Amazon ECS object. Attributes enable you to extend the Amazon ECS data model by adding custom metadata to your resources. For more information, see Attributes in the Amazon Elastic Container Service Developer Guide.

See: newAttribute smart constructor.

Constructors

Attribute'

Fields

targetId :: Maybe Text

The ID of the target. You can specify the short form ID for a resource or the full Amazon Resource Name (ARN).

value :: Maybe Text

The value of the attribute. The value must contain between 1 and 128 characters and may contain letters (uppercase and lowercase), numbers, hyphens, underscores, periods, at signs (@), forward slashes, back slashes, colons, or spaces. The value cannot contain any leading or trailing whitespace.

targetType :: Maybe TargetType

The type of the target with which to attach the attribute. This parameter is required if you use the short form ID for a resource instead of the full ARN.

name :: Text

The name of the attribute. The name must contain between 1 and 128 characters and name may contain letters (uppercase and lowercase), numbers, hyphens, underscores, forward slashes, back slashes, or periods.

newAttribute

Arguments

:: Text
$sel:name:Attribute'

-> Attribute

Create a value of Attribute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetId:Attribute', attribute_targetId - The ID of the target. You can specify the short form ID for a resource or the full Amazon Resource Name (ARN).

$sel:value:Attribute', attribute_value - The value of the attribute. The value must contain between 1 and 128 characters and may contain letters (uppercase and lowercase), numbers, hyphens, underscores, periods, at signs (@), forward slashes, back slashes, colons, or spaces. The value cannot contain any leading or trailing whitespace.

$sel:targetType:Attribute', attribute_targetType - The type of the target with which to attach the attribute. This parameter is required if you use the short form ID for a resource instead of the full ARN.

$sel:name:Attribute', attribute_name - The name of the attribute. The name must contain between 1 and 128 characters and name may contain letters (uppercase and lowercase), numbers, hyphens, underscores, forward slashes, back slashes, or periods.

attribute_targetId :: Lens' Attribute (Maybe Text)

The ID of the target. You can specify the short form ID for a resource or the full Amazon Resource Name (ARN).

attribute_value :: Lens' Attribute (Maybe Text)

The value of the attribute. The value must contain between 1 and 128 characters and may contain letters (uppercase and lowercase), numbers, hyphens, underscores, periods, at signs (@), forward slashes, back slashes, colons, or spaces. The value cannot contain any leading or trailing whitespace.

attribute_targetType :: Lens' Attribute (Maybe TargetType)

The type of the target with which to attach the attribute. This parameter is required if you use the short form ID for a resource instead of the full ARN.

attribute_name :: Lens' Attribute Text

The name of the attribute. The name must contain between 1 and 128 characters and name may contain letters (uppercase and lowercase), numbers, hyphens, underscores, forward slashes, back slashes, or periods.