Network.AWS.ECS.Types.AutoScalingGroupProviderUpdate

Description

Documentation

data AutoScalingGroupProviderUpdate

The details of the Auto Scaling group capacity provider to update.

See: newAutoScalingGroupProviderUpdate smart constructor.

Constructors

AutoScalingGroupProviderUpdate'

Fields

managedScaling :: Maybe ManagedScaling

The managed scaling settings for the Auto Scaling group capacity provider.

managedTerminationProtection :: Maybe ManagedTerminationProtection

The managed termination protection setting to use for the Auto Scaling group capacity provider. This determines whether the Auto Scaling group has managed termination protection.

When using managed termination protection, managed scaling must also be used otherwise managed termination protection will not work.

When managed termination protection is enabled, Amazon ECS prevents the Amazon EC2 instances in an Auto Scaling group that contain tasks from being terminated during a scale-in action. The Auto Scaling group and each instance in the Auto Scaling group must have instance protection from scale-in actions enabled as well. For more information, see Instance Protection in the Auto Scaling User Guide.

When managed termination protection is disabled, your Amazon EC2 instances are not protected from termination when the Auto Scaling group scales in.

newAutoScalingGroupProviderUpdate :: AutoScalingGroupProviderUpdate

Create a value of AutoScalingGroupProviderUpdate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:managedScaling:AutoScalingGroupProviderUpdate', autoScalingGroupProviderUpdate_managedScaling - The managed scaling settings for the Auto Scaling group capacity provider.

$sel:managedTerminationProtection:AutoScalingGroupProviderUpdate', autoScalingGroupProviderUpdate_managedTerminationProtection - The managed termination protection setting to use for the Auto Scaling group capacity provider. This determines whether the Auto Scaling group has managed termination protection.

When using managed termination protection, managed scaling must also be used otherwise managed termination protection will not work.

When managed termination protection is enabled, Amazon ECS prevents the Amazon EC2 instances in an Auto Scaling group that contain tasks from being terminated during a scale-in action. The Auto Scaling group and each instance in the Auto Scaling group must have instance protection from scale-in actions enabled as well. For more information, see Instance Protection in the Auto Scaling User Guide.

When managed termination protection is disabled, your Amazon EC2 instances are not protected from termination when the Auto Scaling group scales in.

autoScalingGroupProviderUpdate_managedScaling :: Lens' AutoScalingGroupProviderUpdate (Maybe ManagedScaling)

The managed scaling settings for the Auto Scaling group capacity provider.

autoScalingGroupProviderUpdate_managedTerminationProtection :: Lens' AutoScalingGroupProviderUpdate (Maybe ManagedTerminationProtection)

The managed termination protection setting to use for the Auto Scaling group capacity provider. This determines whether the Auto Scaling group has managed termination protection.

When using managed termination protection, managed scaling must also be used otherwise managed termination protection will not work.

When managed termination protection is enabled, Amazon ECS prevents the Amazon EC2 instances in an Auto Scaling group that contain tasks from being terminated during a scale-in action. The Auto Scaling group and each instance in the Auto Scaling group must have instance protection from scale-in actions enabled as well. For more information, see Instance Protection in the Auto Scaling User Guide.

When managed termination protection is disabled, your Amazon EC2 instances are not protected from termination when the Auto Scaling group scales in.