Network.AWS.ECS.Types.AwsVpcConfiguration

Description

Documentation

data AwsVpcConfiguration

An object representing the networking details for a task or service.

See: newAwsVpcConfiguration smart constructor.

Constructors

AwsVpcConfiguration'

Fields

assignPublicIp :: Maybe AssignPublicIp

Whether the task's elastic network interface receives a public IP address. The default value is DISABLED.

securityGroups :: Maybe [Text]

The IDs of the security groups associated with the task or service. If you do not specify a security group, the default security group for the VPC is used. There is a limit of 5 security groups that can be specified per AwsVpcConfiguration.

All specified security groups must be from the same VPC.

subnets :: [Text]

The IDs of the subnets associated with the task or service. There is a limit of 16 subnets that can be specified per AwsVpcConfiguration.

All specified subnets must be from the same VPC.

newAwsVpcConfiguration :: AwsVpcConfiguration

Create a value of AwsVpcConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:assignPublicIp:AwsVpcConfiguration', awsVpcConfiguration_assignPublicIp - Whether the task's elastic network interface receives a public IP address. The default value is DISABLED.

$sel:securityGroups:AwsVpcConfiguration', awsVpcConfiguration_securityGroups - The IDs of the security groups associated with the task or service. If you do not specify a security group, the default security group for the VPC is used. There is a limit of 5 security groups that can be specified per AwsVpcConfiguration.

All specified security groups must be from the same VPC.

$sel:subnets:AwsVpcConfiguration', awsVpcConfiguration_subnets - The IDs of the subnets associated with the task or service. There is a limit of 16 subnets that can be specified per AwsVpcConfiguration.

All specified subnets must be from the same VPC.

awsVpcConfiguration_assignPublicIp :: Lens' AwsVpcConfiguration (Maybe AssignPublicIp)

Whether the task's elastic network interface receives a public IP address. The default value is DISABLED.

awsVpcConfiguration_securityGroups :: Lens' AwsVpcConfiguration (Maybe [Text])

The IDs of the security groups associated with the task or service. If you do not specify a security group, the default security group for the VPC is used. There is a limit of 5 security groups that can be specified per AwsVpcConfiguration.

All specified security groups must be from the same VPC.

awsVpcConfiguration_subnets :: Lens' AwsVpcConfiguration [Text]

The IDs of the subnets associated with the task or service. There is a limit of 16 subnets that can be specified per AwsVpcConfiguration.

All specified subnets must be from the same VPC.