Network.AWS.ECS.Types.Cluster

Description

Documentation

data Cluster

A regional grouping of one or more container instances on which you can run task requests. Each account receives a default cluster the first time you use the Amazon ECS service, but you may also create other clusters. Clusters may contain more than one instance type simultaneously.

See: newCluster smart constructor.

Constructors

Cluster'

Fields

clusterArn :: Maybe Text

The Amazon Resource Name (ARN) that identifies the cluster. The ARN contains the arn:aws:ecs namespace, followed by the Region of the cluster, the Amazon Web Services account ID of the cluster owner, the cluster namespace, and then the cluster name. For example, arn:aws:ecs:region:012345678910:cluster/test.

status :: Maybe Text

The status of the cluster. The following are the possible states that will be returned.

ACTIVE

The cluster is ready to accept tasks and if applicable you can register container instances with the cluster.

PROVISIONING

The cluster has capacity providers associated with it and the resources needed for the capacity provider are being created.

DEPROVISIONING

The cluster has capacity providers associated with it and the resources needed for the capacity provider are being deleted.

FAILED

The cluster has capacity providers associated with it and the resources needed for the capacity provider have failed to create.

INACTIVE

The cluster has been deleted. Clusters with an INACTIVE status may remain discoverable in your account for a period of time. However, this behavior is subject to change in the future, so you should not rely on INACTIVE clusters persisting.

activeServicesCount :: Maybe Int

The number of services that are running on the cluster in an ACTIVE state. You can view these services with ListServices.

registeredContainerInstancesCount :: Maybe Int

The number of container instances registered into the cluster. This includes container instances in both ACTIVE and DRAINING status.

configuration :: Maybe ClusterConfiguration

The execute command configuration for the cluster.

statistics :: Maybe [KeyValuePair]

Additional information about your clusters that are separated by launch type, including:

runningEC2TasksCount

RunningFargateTasksCount

pendingEC2TasksCount

pendingFargateTasksCount

activeEC2ServiceCount

activeFargateServiceCount

drainingEC2ServiceCount

drainingFargateServiceCount

defaultCapacityProviderStrategy :: Maybe [CapacityProviderStrategyItem]

The default capacity provider strategy for the cluster. When services or tasks are run in the cluster with no launch type or capacity provider strategy specified, the default capacity provider strategy is used.

pendingTasksCount :: Maybe Int

The number of tasks in the cluster that are in the PENDING state.

tags :: Maybe [Tag]

The metadata that you apply to the cluster to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define.

The following basic restrictions apply to tags:

Maximum number of tags per resource - 50

For each resource, each tag key must be unique, and each tag key can have only one value.

Maximum key length - 128 Unicode characters in UTF-8

Maximum value length - 256 Unicode characters in UTF-8

If your tagging schema is used across multiple services and resources, remember that other services may have restrictions on allowed characters. Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following characters: + - = . _ : / @.

Tag keys and values are case-sensitive.

Do not use aws:, AWS:, or any upper or lowercase combination of such as a prefix for either keys or values as it is reserved for Amazon Web Services use. You cannot edit or delete tag keys or values with this prefix. Tags with this prefix do not count against your tags per resource limit.

capacityProviders :: Maybe [Text]

The capacity providers associated with the cluster.

attachmentsStatus :: Maybe Text

The status of the capacity providers associated with the cluster. The following are the states that will be returned:

UPDATE_IN_PROGRESS

The available capacity providers for the cluster are updating. This occurs when the Auto Scaling plan is provisioning or deprovisioning.

UPDATE_COMPLETE

The capacity providers have successfully updated.

UPDATE_FAILED

The capacity provider updates failed.

clusterName :: Maybe Text

A user-generated string that you use to identify your cluster.

settings :: Maybe [ClusterSetting]

The settings for the cluster. This parameter indicates whether CloudWatch Container Insights is enabled or disabled for a cluster.

runningTasksCount :: Maybe Int

The number of tasks in the cluster that are in the RUNNING state.

attachments :: Maybe [Attachment]

The resources attached to a cluster. When using a capacity provider with a cluster, the Auto Scaling plan that is created will be returned as a cluster attachment.

newCluster :: Cluster

Create a value of Cluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clusterArn:Cluster', cluster_clusterArn - The Amazon Resource Name (ARN) that identifies the cluster. The ARN contains the arn:aws:ecs namespace, followed by the Region of the cluster, the Amazon Web Services account ID of the cluster owner, the cluster namespace, and then the cluster name. For example, arn:aws:ecs:region:012345678910:cluster/test.

$sel:status:Cluster', cluster_status - The status of the cluster. The following are the possible states that will be returned.

ACTIVE

The cluster is ready to accept tasks and if applicable you can register container instances with the cluster.

PROVISIONING

The cluster has capacity providers associated with it and the resources needed for the capacity provider are being created.

DEPROVISIONING

The cluster has capacity providers associated with it and the resources needed for the capacity provider are being deleted.

FAILED

The cluster has capacity providers associated with it and the resources needed for the capacity provider have failed to create.

INACTIVE

The cluster has been deleted. Clusters with an INACTIVE status may remain discoverable in your account for a period of time. However, this behavior is subject to change in the future, so you should not rely on INACTIVE clusters persisting.

$sel:activeServicesCount:Cluster', cluster_activeServicesCount - The number of services that are running on the cluster in an ACTIVE state. You can view these services with ListServices.

$sel:registeredContainerInstancesCount:Cluster', cluster_registeredContainerInstancesCount - The number of container instances registered into the cluster. This includes container instances in both ACTIVE and DRAINING status.

$sel:configuration:Cluster', cluster_configuration - The execute command configuration for the cluster.

$sel:statistics:Cluster', cluster_statistics - Additional information about your clusters that are separated by launch type, including:

runningEC2TasksCount

RunningFargateTasksCount

pendingEC2TasksCount

pendingFargateTasksCount

activeEC2ServiceCount

activeFargateServiceCount

drainingEC2ServiceCount

drainingFargateServiceCount

$sel:defaultCapacityProviderStrategy:Cluster', cluster_defaultCapacityProviderStrategy - The default capacity provider strategy for the cluster. When services or tasks are run in the cluster with no launch type or capacity provider strategy specified, the default capacity provider strategy is used.

$sel:pendingTasksCount:Cluster', cluster_pendingTasksCount - The number of tasks in the cluster that are in the PENDING state.

$sel:tags:Cluster', cluster_tags - The metadata that you apply to the cluster to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define.

The following basic restrictions apply to tags:

Maximum number of tags per resource - 50

For each resource, each tag key must be unique, and each tag key can have only one value.

Maximum key length - 128 Unicode characters in UTF-8

Maximum value length - 256 Unicode characters in UTF-8

If your tagging schema is used across multiple services and resources, remember that other services may have restrictions on allowed characters. Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following characters: + - = . _ : / @.

Tag keys and values are case-sensitive.

Do not use aws:, AWS:, or any upper or lowercase combination of such as a prefix for either keys or values as it is reserved for Amazon Web Services use. You cannot edit or delete tag keys or values with this prefix. Tags with this prefix do not count against your tags per resource limit.

$sel:capacityProviders:Cluster', cluster_capacityProviders - The capacity providers associated with the cluster.

$sel:attachmentsStatus:Cluster', cluster_attachmentsStatus - The status of the capacity providers associated with the cluster. The following are the states that will be returned:

UPDATE_IN_PROGRESS

The available capacity providers for the cluster are updating. This occurs when the Auto Scaling plan is provisioning or deprovisioning.

UPDATE_COMPLETE

The capacity providers have successfully updated.

UPDATE_FAILED

The capacity provider updates failed.

$sel:clusterName:Cluster', cluster_clusterName - A user-generated string that you use to identify your cluster.

$sel:settings:Cluster', cluster_settings - The settings for the cluster. This parameter indicates whether CloudWatch Container Insights is enabled or disabled for a cluster.

$sel:runningTasksCount:Cluster', cluster_runningTasksCount - The number of tasks in the cluster that are in the RUNNING state.

$sel:attachments:Cluster', cluster_attachments - The resources attached to a cluster. When using a capacity provider with a cluster, the Auto Scaling plan that is created will be returned as a cluster attachment.

cluster_clusterArn :: Lens' Cluster (Maybe Text)

The Amazon Resource Name (ARN) that identifies the cluster. The ARN contains the arn:aws:ecs namespace, followed by the Region of the cluster, the Amazon Web Services account ID of the cluster owner, the cluster namespace, and then the cluster name. For example, arn:aws:ecs:region:012345678910:cluster/test.

cluster_status :: Lens' Cluster (Maybe Text)

The status of the cluster. The following are the possible states that will be returned.

ACTIVE

The cluster is ready to accept tasks and if applicable you can register container instances with the cluster.

PROVISIONING

The cluster has capacity providers associated with it and the resources needed for the capacity provider are being created.

DEPROVISIONING

The cluster has capacity providers associated with it and the resources needed for the capacity provider are being deleted.

FAILED

The cluster has capacity providers associated with it and the resources needed for the capacity provider have failed to create.

INACTIVE

The cluster has been deleted. Clusters with an INACTIVE status may remain discoverable in your account for a period of time. However, this behavior is subject to change in the future, so you should not rely on INACTIVE clusters persisting.

cluster_activeServicesCount :: Lens' Cluster (Maybe Int)

The number of services that are running on the cluster in an ACTIVE state. You can view these services with ListServices.

cluster_registeredContainerInstancesCount :: Lens' Cluster (Maybe Int)

The number of container instances registered into the cluster. This includes container instances in both ACTIVE and DRAINING status.

cluster_configuration :: Lens' Cluster (Maybe ClusterConfiguration)

The execute command configuration for the cluster.

cluster_statistics :: Lens' Cluster (Maybe [KeyValuePair])

Additional information about your clusters that are separated by launch type, including:

runningEC2TasksCount

RunningFargateTasksCount

pendingEC2TasksCount

pendingFargateTasksCount

activeEC2ServiceCount

activeFargateServiceCount

drainingEC2ServiceCount

drainingFargateServiceCount

cluster_defaultCapacityProviderStrategy :: Lens' Cluster (Maybe [CapacityProviderStrategyItem])

The default capacity provider strategy for the cluster. When services or tasks are run in the cluster with no launch type or capacity provider strategy specified, the default capacity provider strategy is used.

cluster_pendingTasksCount :: Lens' Cluster (Maybe Int)

The number of tasks in the cluster that are in the PENDING state.

cluster_tags :: Lens' Cluster (Maybe [Tag])

The metadata that you apply to the cluster to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define.

The following basic restrictions apply to tags:

Maximum number of tags per resource - 50

For each resource, each tag key must be unique, and each tag key can have only one value.

Maximum key length - 128 Unicode characters in UTF-8

Maximum value length - 256 Unicode characters in UTF-8

If your tagging schema is used across multiple services and resources, remember that other services may have restrictions on allowed characters. Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following characters: + - = . _ : / @.

Tag keys and values are case-sensitive.

Do not use aws:, AWS:, or any upper or lowercase combination of such as a prefix for either keys or values as it is reserved for Amazon Web Services use. You cannot edit or delete tag keys or values with this prefix. Tags with this prefix do not count against your tags per resource limit.

cluster_capacityProviders :: Lens' Cluster (Maybe [Text])

The capacity providers associated with the cluster.

cluster_attachmentsStatus :: Lens' Cluster (Maybe Text)

The status of the capacity providers associated with the cluster. The following are the states that will be returned:

UPDATE_IN_PROGRESS

The available capacity providers for the cluster are updating. This occurs when the Auto Scaling plan is provisioning or deprovisioning.

UPDATE_COMPLETE

The capacity providers have successfully updated.

UPDATE_FAILED

The capacity provider updates failed.

cluster_clusterName :: Lens' Cluster (Maybe Text)

A user-generated string that you use to identify your cluster.

cluster_settings :: Lens' Cluster (Maybe [ClusterSetting])

The settings for the cluster. This parameter indicates whether CloudWatch Container Insights is enabled or disabled for a cluster.

cluster_runningTasksCount :: Lens' Cluster (Maybe Int)

The number of tasks in the cluster that are in the RUNNING state.

cluster_attachments :: Lens' Cluster (Maybe [Attachment])

The resources attached to a cluster. When using a capacity provider with a cluster, the Auto Scaling plan that is created will be returned as a cluster attachment.