Network.AWS.ECS.Types.ClusterSetting

Description

Documentation

data ClusterSetting

The settings to use when creating a cluster. This parameter is used to enable CloudWatch Container Insights for a cluster.

See: newClusterSetting smart constructor.

Constructors

ClusterSetting'

Fields

name :: Maybe ClusterSettingName

The name of the cluster setting. The only supported value is containerInsights.

value :: Maybe Text

The value to set for the cluster setting. The supported values are enabled and disabled. If enabled is specified, CloudWatch Container Insights will be enabled for the cluster, otherwise it will be disabled unless the containerInsights account setting is enabled. If a cluster value is specified, it will override the containerInsights value set with PutAccountSetting or PutAccountSettingDefault.

newClusterSetting :: ClusterSetting

Create a value of ClusterSetting with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:ClusterSetting', clusterSetting_name - The name of the cluster setting. The only supported value is containerInsights.

$sel:value:ClusterSetting', clusterSetting_value - The value to set for the cluster setting. The supported values are enabled and disabled. If enabled is specified, CloudWatch Container Insights will be enabled for the cluster, otherwise it will be disabled unless the containerInsights account setting is enabled. If a cluster value is specified, it will override the containerInsights value set with PutAccountSetting or PutAccountSettingDefault.

clusterSetting_name :: Lens' ClusterSetting (Maybe ClusterSettingName)

The name of the cluster setting. The only supported value is containerInsights.

clusterSetting_value :: Lens' ClusterSetting (Maybe Text)

The value to set for the cluster setting. The supported values are enabled and disabled. If enabled is specified, CloudWatch Container Insights will be enabled for the cluster, otherwise it will be disabled unless the containerInsights account setting is enabled. If a cluster value is specified, it will override the containerInsights value set with PutAccountSetting or PutAccountSettingDefault.