Network.AWS.ECS.Types.Container

Description

Documentation

data Container

A Docker container that is part of a task.

See: newContainer smart constructor.

Constructors

Container'

Fields

imageDigest :: Maybe Text

The container image manifest digest.

The imageDigest is only returned if the container is using an image hosted in Amazon ECR, otherwise it is omitted.

gpuIds :: Maybe [Text]

The IDs of each GPU assigned to the container.

memoryReservation :: Maybe Text

The soft limit (in MiB) of memory set for the container.

memory :: Maybe Text

The hard limit (in MiB) of memory set for the container.

runtimeId :: Maybe Text

The ID of the Docker container.

exitCode :: Maybe Int

The exit code returned from the container.

containerArn :: Maybe Text

The Amazon Resource Name (ARN) of the container.

name :: Maybe Text

The name of the container.

image :: Maybe Text

The image used for the container.

managedAgents :: Maybe [ManagedAgent]

The details of any Amazon ECS managed agents associated with the container.

networkBindings :: Maybe [NetworkBinding]

The network bindings associated with the container.

reason :: Maybe Text

A short (255 max characters) human-readable string to provide additional details about a running or stopped container.

lastStatus :: Maybe Text

The last known status of the container.

cpu :: Maybe Text

The number of CPU units set for the container. The value will be 0 if no value was specified in the container definition when the task definition was registered.

networkInterfaces :: Maybe [NetworkInterface]

The network interfaces associated with the container.

healthStatus :: Maybe HealthStatus

The health status of the container. If health checks are not configured for this container in its task definition, then it reports the health status as UNKNOWN.

taskArn :: Maybe Text

The ARN of the task.

newContainer :: Container

Create a value of Container with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:imageDigest:Container', container_imageDigest - The container image manifest digest.

The imageDigest is only returned if the container is using an image hosted in Amazon ECR, otherwise it is omitted.

$sel:gpuIds:Container', container_gpuIds - The IDs of each GPU assigned to the container.

$sel:memoryReservation:Container', container_memoryReservation - The soft limit (in MiB) of memory set for the container.

$sel:memory:Container', container_memory - The hard limit (in MiB) of memory set for the container.

$sel:runtimeId:Container', container_runtimeId - The ID of the Docker container.

$sel:exitCode:Container', container_exitCode - The exit code returned from the container.

$sel:containerArn:Container', container_containerArn - The Amazon Resource Name (ARN) of the container.

$sel:name:Container', container_name - The name of the container.

$sel:image:Container', container_image - The image used for the container.

$sel:managedAgents:Container', container_managedAgents - The details of any Amazon ECS managed agents associated with the container.

$sel:networkBindings:Container', container_networkBindings - The network bindings associated with the container.

$sel:reason:Container', container_reason - A short (255 max characters) human-readable string to provide additional details about a running or stopped container.

$sel:lastStatus:Container', container_lastStatus - The last known status of the container.

$sel:cpu:Container', container_cpu - The number of CPU units set for the container. The value will be 0 if no value was specified in the container definition when the task definition was registered.

$sel:networkInterfaces:Container', container_networkInterfaces - The network interfaces associated with the container.

$sel:healthStatus:Container', container_healthStatus - The health status of the container. If health checks are not configured for this container in its task definition, then it reports the health status as UNKNOWN.

$sel:taskArn:Container', container_taskArn - The ARN of the task.

container_imageDigest :: Lens' Container (Maybe Text)

The container image manifest digest.

The imageDigest is only returned if the container is using an image hosted in Amazon ECR, otherwise it is omitted.

container_gpuIds :: Lens' Container (Maybe [Text])

The IDs of each GPU assigned to the container.

container_memoryReservation :: Lens' Container (Maybe Text)

The soft limit (in MiB) of memory set for the container.

container_memory :: Lens' Container (Maybe Text)

The hard limit (in MiB) of memory set for the container.

container_runtimeId :: Lens' Container (Maybe Text)

The ID of the Docker container.

container_exitCode :: Lens' Container (Maybe Int)

The exit code returned from the container.

container_containerArn :: Lens' Container (Maybe Text)

The Amazon Resource Name (ARN) of the container.

container_name :: Lens' Container (Maybe Text)

The name of the container.

container_image :: Lens' Container (Maybe Text)

The image used for the container.

container_managedAgents :: Lens' Container (Maybe [ManagedAgent])

The details of any Amazon ECS managed agents associated with the container.

container_networkBindings :: Lens' Container (Maybe [NetworkBinding])

The network bindings associated with the container.

container_reason :: Lens' Container (Maybe Text)

A short (255 max characters) human-readable string to provide additional details about a running or stopped container.

container_lastStatus :: Lens' Container (Maybe Text)

The last known status of the container.

container_cpu :: Lens' Container (Maybe Text)

The number of CPU units set for the container. The value will be 0 if no value was specified in the container definition when the task definition was registered.

container_networkInterfaces :: Lens' Container (Maybe [NetworkInterface])

The network interfaces associated with the container.

container_healthStatus :: Lens' Container (Maybe HealthStatus)

The health status of the container. If health checks are not configured for this container in its task definition, then it reports the health status as UNKNOWN.

container_taskArn :: Lens' Container (Maybe Text)

The ARN of the task.