Network.AWS.ECS.Types.ContainerService

Description

Documentation

data ContainerService

Details on a service within a cluster

See: newContainerService smart constructor.

Constructors

ContainerService'

Fields

clusterArn :: Maybe Text

The Amazon Resource Name (ARN) of the cluster that hosts the service.

taskSets :: Maybe [TaskSet]

Information about a set of Amazon ECS tasks in either an CodeDeploy or an EXTERNAL deployment. An Amazon ECS task set includes details such as the desired number of tasks, how many tasks are running, and whether the task set serves production traffic.

runningCount :: Maybe Int

The number of tasks in the cluster that are in the RUNNING state.

status :: Maybe Text

The status of the service. The valid values are ACTIVE, DRAINING, or INACTIVE.

roleArn :: Maybe Text

The ARN of the IAM role associated with the service that allows the Amazon ECS container agent to register container instances with an Elastic Load Balancing load balancer.

deploymentConfiguration :: Maybe DeploymentConfiguration

Optional deployment parameters that control how many tasks run during the deployment and the ordering of stopping and starting tasks.

capacityProviderStrategy :: Maybe [CapacityProviderStrategyItem]

The capacity provider strategy the service is using. When using the DescribeServices API, this field is omitted if the service was created using a launch type.

networkConfiguration :: Maybe NetworkConfiguration

The VPC subnet and security group configuration for tasks that receive their own elastic network interface by using the awsvpc networking mode.

desiredCount :: Maybe Int

The desired number of instantiations of the task definition to keep running on the service. This value is specified when the service is created with CreateService, and it can be modified with UpdateService.

enableECSManagedTags :: Maybe Bool

Specifies whether to enable Amazon ECS managed tags for the tasks in the service. For more information, see Tagging Your Amazon ECS Resources in the Amazon Elastic Container Service Developer Guide.

deploymentController :: Maybe DeploymentController

The deployment controller type the service is using. When using the DescribeServices API, this field is omitted if the service is using the ECS deployment controller type.

launchType :: Maybe LaunchType

The launch type the service is using. When using the DescribeServices API, this field is omitted if the service was created using a capacity provider strategy.

createdAt :: Maybe POSIX

The Unix timestamp for when the service was created.

platformVersion :: Maybe Text

The platform version on which to run your service. A platform version is only specified for tasks hosted on Fargate. If one is not specified, the LATEST platform version is used by default. For more information, see Fargate Platform Versions in the Amazon Elastic Container Service Developer Guide.

deployments :: Maybe [Deployment]

The current state of deployments for the service.

placementStrategy :: Maybe [PlacementStrategy]

The placement strategy that determines how tasks for the service are placed.

serviceName :: Maybe Text

The name of your service. Up to 255 letters (uppercase and lowercase), numbers, underscores, and hyphens are allowed. Service names must be unique within a cluster, but you can have similarly named services in multiple clusters within a Region or across multiple Regions.

placementConstraints :: Maybe [PlacementConstraint]

The placement constraints for the tasks in the service.

events :: Maybe [ServiceEvent]

The event stream for your service. A maximum of 100 of the latest events are displayed.

pendingCount :: Maybe Int

The number of tasks in the cluster that are in the PENDING state.

enableExecuteCommand :: Maybe Bool

Whether or not the execute command functionality is enabled for the service. If true, the execute command functionality is enabled for all containers in tasks as part of the service.

loadBalancers :: Maybe [LoadBalancer]

A list of Elastic Load Balancing load balancer objects, containing the load balancer name, the container name (as it appears in a container definition), and the container port to access from the load balancer.

tags :: Maybe [Tag]

The metadata that you apply to the service to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define.

The following basic restrictions apply to tags:

Maximum number of tags per resource - 50

For each resource, each tag key must be unique, and each tag key can have only one value.

Maximum key length - 128 Unicode characters in UTF-8

Maximum value length - 256 Unicode characters in UTF-8

If your tagging schema is used across multiple services and resources, remember that other services may have restrictions on allowed characters. Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following characters: + - = . _ : / @.

Tag keys and values are case-sensitive.

Do not use aws:, AWS:, or any upper or lowercase combination of such as a prefix for either keys or values as it is reserved for Amazon Web Services use. You cannot edit or delete tag keys or values with this prefix. Tags with this prefix do not count against your tags per resource limit.

healthCheckGracePeriodSeconds :: Maybe Int

The period of time, in seconds, that the Amazon ECS service scheduler ignores unhealthy Elastic Load Balancing target health checks after a task has first started.

serviceRegistries :: Maybe [ServiceRegistry]

The details of the service discovery registries to assign to this service. For more information, see Service Discovery.

createdBy :: Maybe Text

The principal that created the service.

schedulingStrategy :: Maybe SchedulingStrategy

The scheduling strategy to use for the service. For more information, see Services.

There are two service scheduler strategies available:

REPLICA-The replica scheduling strategy places and maintains the desired number of tasks across your cluster. By default, the service scheduler spreads tasks across Availability Zones. You can use task placement strategies and constraints to customize task placement decisions.

DAEMON-The daemon scheduling strategy deploys exactly one task on each active container instance that meets all of the task placement constraints that you specify in your cluster. The service scheduler also evaluates the task placement constraints for running tasks and will stop tasks that do not meet the placement constraints.

Fargate tasks do not support the DAEMON scheduling strategy.

taskDefinition :: Maybe Text

The task definition to use for tasks in the service. This value is specified when the service is created with CreateService, and it can be modified with UpdateService.

serviceArn :: Maybe Text

The ARN that identifies the service. The ARN contains the arn:aws:ecs namespace, followed by the Region of the service, the Amazon Web Services account ID of the service owner, the service namespace, and then the service name. For example, arn:aws:ecs:region:012345678910:service/my-service.

propagateTags :: Maybe PropagateTags

Specifies whether to propagate the tags from the task definition or the service to the task. If no value is specified, the tags are not propagated.

newContainerService :: ContainerService

Create a value of ContainerService with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clusterArn:ContainerService', containerService_clusterArn - The Amazon Resource Name (ARN) of the cluster that hosts the service.

$sel:taskSets:ContainerService', containerService_taskSets - Information about a set of Amazon ECS tasks in either an CodeDeploy or an EXTERNAL deployment. An Amazon ECS task set includes details such as the desired number of tasks, how many tasks are running, and whether the task set serves production traffic.

$sel:runningCount:ContainerService', containerService_runningCount - The number of tasks in the cluster that are in the RUNNING state.

$sel:status:ContainerService', containerService_status - The status of the service. The valid values are ACTIVE, DRAINING, or INACTIVE.

$sel:roleArn:ContainerService', containerService_roleArn - The ARN of the IAM role associated with the service that allows the Amazon ECS container agent to register container instances with an Elastic Load Balancing load balancer.

$sel:deploymentConfiguration:ContainerService', containerService_deploymentConfiguration - Optional deployment parameters that control how many tasks run during the deployment and the ordering of stopping and starting tasks.

$sel:capacityProviderStrategy:ContainerService', containerService_capacityProviderStrategy - The capacity provider strategy the service is using. When using the DescribeServices API, this field is omitted if the service was created using a launch type.

$sel:networkConfiguration:ContainerService', containerService_networkConfiguration - The VPC subnet and security group configuration for tasks that receive their own elastic network interface by using the awsvpc networking mode.

$sel:desiredCount:ContainerService', containerService_desiredCount - The desired number of instantiations of the task definition to keep running on the service. This value is specified when the service is created with CreateService, and it can be modified with UpdateService.

$sel:enableECSManagedTags:ContainerService', containerService_enableECSManagedTags - Specifies whether to enable Amazon ECS managed tags for the tasks in the service. For more information, see Tagging Your Amazon ECS Resources in the Amazon Elastic Container Service Developer Guide.

$sel:deploymentController:ContainerService', containerService_deploymentController - The deployment controller type the service is using. When using the DescribeServices API, this field is omitted if the service is using the ECS deployment controller type.

$sel:launchType:ContainerService', containerService_launchType - The launch type the service is using. When using the DescribeServices API, this field is omitted if the service was created using a capacity provider strategy.

$sel:createdAt:ContainerService', containerService_createdAt - The Unix timestamp for when the service was created.

$sel:platformVersion:ContainerService', containerService_platformVersion - The platform version on which to run your service. A platform version is only specified for tasks hosted on Fargate. If one is not specified, the LATEST platform version is used by default. For more information, see Fargate Platform Versions in the Amazon Elastic Container Service Developer Guide.

$sel:deployments:ContainerService', containerService_deployments - The current state of deployments for the service.

$sel:placementStrategy:ContainerService', containerService_placementStrategy - The placement strategy that determines how tasks for the service are placed.

$sel:serviceName:ContainerService', containerService_serviceName - The name of your service. Up to 255 letters (uppercase and lowercase), numbers, underscores, and hyphens are allowed. Service names must be unique within a cluster, but you can have similarly named services in multiple clusters within a Region or across multiple Regions.

$sel:placementConstraints:ContainerService', containerService_placementConstraints - The placement constraints for the tasks in the service.

$sel:events:ContainerService', containerService_events - The event stream for your service. A maximum of 100 of the latest events are displayed.

$sel:pendingCount:ContainerService', containerService_pendingCount - The number of tasks in the cluster that are in the PENDING state.

$sel:enableExecuteCommand:ContainerService', containerService_enableExecuteCommand - Whether or not the execute command functionality is enabled for the service. If true, the execute command functionality is enabled for all containers in tasks as part of the service.

$sel:loadBalancers:ContainerService', containerService_loadBalancers - A list of Elastic Load Balancing load balancer objects, containing the load balancer name, the container name (as it appears in a container definition), and the container port to access from the load balancer.

$sel:tags:ContainerService', containerService_tags - The metadata that you apply to the service to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define.

The following basic restrictions apply to tags:

Maximum number of tags per resource - 50

For each resource, each tag key must be unique, and each tag key can have only one value.

Maximum key length - 128 Unicode characters in UTF-8

Maximum value length - 256 Unicode characters in UTF-8

If your tagging schema is used across multiple services and resources, remember that other services may have restrictions on allowed characters. Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following characters: + - = . _ : / @.

Tag keys and values are case-sensitive.

Do not use aws:, AWS:, or any upper or lowercase combination of such as a prefix for either keys or values as it is reserved for Amazon Web Services use. You cannot edit or delete tag keys or values with this prefix. Tags with this prefix do not count against your tags per resource limit.

$sel:healthCheckGracePeriodSeconds:ContainerService', containerService_healthCheckGracePeriodSeconds - The period of time, in seconds, that the Amazon ECS service scheduler ignores unhealthy Elastic Load Balancing target health checks after a task has first started.

$sel:serviceRegistries:ContainerService', containerService_serviceRegistries - The details of the service discovery registries to assign to this service. For more information, see Service Discovery.

$sel:createdBy:ContainerService', containerService_createdBy - The principal that created the service.

$sel:schedulingStrategy:ContainerService', containerService_schedulingStrategy - The scheduling strategy to use for the service. For more information, see Services.

There are two service scheduler strategies available:

REPLICA-The replica scheduling strategy places and maintains the desired number of tasks across your cluster. By default, the service scheduler spreads tasks across Availability Zones. You can use task placement strategies and constraints to customize task placement decisions.

DAEMON-The daemon scheduling strategy deploys exactly one task on each active container instance that meets all of the task placement constraints that you specify in your cluster. The service scheduler also evaluates the task placement constraints for running tasks and will stop tasks that do not meet the placement constraints.

Fargate tasks do not support the DAEMON scheduling strategy.

$sel:taskDefinition:ContainerService', containerService_taskDefinition - The task definition to use for tasks in the service. This value is specified when the service is created with CreateService, and it can be modified with UpdateService.

$sel:serviceArn:ContainerService', containerService_serviceArn - The ARN that identifies the service. The ARN contains the arn:aws:ecs namespace, followed by the Region of the service, the Amazon Web Services account ID of the service owner, the service namespace, and then the service name. For example, arn:aws:ecs:region:012345678910:service/my-service.

$sel:propagateTags:ContainerService', containerService_propagateTags - Specifies whether to propagate the tags from the task definition or the service to the task. If no value is specified, the tags are not propagated.

containerService_clusterArn :: Lens' ContainerService (Maybe Text)

The Amazon Resource Name (ARN) of the cluster that hosts the service.

containerService_taskSets :: Lens' ContainerService (Maybe [TaskSet])

Information about a set of Amazon ECS tasks in either an CodeDeploy or an EXTERNAL deployment. An Amazon ECS task set includes details such as the desired number of tasks, how many tasks are running, and whether the task set serves production traffic.

containerService_runningCount :: Lens' ContainerService (Maybe Int)

The number of tasks in the cluster that are in the RUNNING state.

containerService_status :: Lens' ContainerService (Maybe Text)

The status of the service. The valid values are ACTIVE, DRAINING, or INACTIVE.

containerService_roleArn :: Lens' ContainerService (Maybe Text)

The ARN of the IAM role associated with the service that allows the Amazon ECS container agent to register container instances with an Elastic Load Balancing load balancer.

containerService_deploymentConfiguration :: Lens' ContainerService (Maybe DeploymentConfiguration)

Optional deployment parameters that control how many tasks run during the deployment and the ordering of stopping and starting tasks.

containerService_capacityProviderStrategy :: Lens' ContainerService (Maybe [CapacityProviderStrategyItem])

The capacity provider strategy the service is using. When using the DescribeServices API, this field is omitted if the service was created using a launch type.

containerService_networkConfiguration :: Lens' ContainerService (Maybe NetworkConfiguration)

The VPC subnet and security group configuration for tasks that receive their own elastic network interface by using the awsvpc networking mode.

containerService_desiredCount :: Lens' ContainerService (Maybe Int)

The desired number of instantiations of the task definition to keep running on the service. This value is specified when the service is created with CreateService, and it can be modified with UpdateService.

containerService_enableECSManagedTags :: Lens' ContainerService (Maybe Bool)

Specifies whether to enable Amazon ECS managed tags for the tasks in the service. For more information, see Tagging Your Amazon ECS Resources in the Amazon Elastic Container Service Developer Guide.

containerService_deploymentController :: Lens' ContainerService (Maybe DeploymentController)

The deployment controller type the service is using. When using the DescribeServices API, this field is omitted if the service is using the ECS deployment controller type.

containerService_launchType :: Lens' ContainerService (Maybe LaunchType)

The launch type the service is using. When using the DescribeServices API, this field is omitted if the service was created using a capacity provider strategy.

containerService_createdAt :: Lens' ContainerService (Maybe UTCTime)

The Unix timestamp for when the service was created.

containerService_platformVersion :: Lens' ContainerService (Maybe Text)

The platform version on which to run your service. A platform version is only specified for tasks hosted on Fargate. If one is not specified, the LATEST platform version is used by default. For more information, see Fargate Platform Versions in the Amazon Elastic Container Service Developer Guide.

containerService_deployments :: Lens' ContainerService (Maybe [Deployment])

The current state of deployments for the service.

containerService_placementStrategy :: Lens' ContainerService (Maybe [PlacementStrategy])

The placement strategy that determines how tasks for the service are placed.

containerService_serviceName :: Lens' ContainerService (Maybe Text)

The name of your service. Up to 255 letters (uppercase and lowercase), numbers, underscores, and hyphens are allowed. Service names must be unique within a cluster, but you can have similarly named services in multiple clusters within a Region or across multiple Regions.

containerService_placementConstraints :: Lens' ContainerService (Maybe [PlacementConstraint])

The placement constraints for the tasks in the service.

containerService_events :: Lens' ContainerService (Maybe [ServiceEvent])

The event stream for your service. A maximum of 100 of the latest events are displayed.

containerService_pendingCount :: Lens' ContainerService (Maybe Int)

The number of tasks in the cluster that are in the PENDING state.

containerService_enableExecuteCommand :: Lens' ContainerService (Maybe Bool)

Whether or not the execute command functionality is enabled for the service. If true, the execute command functionality is enabled for all containers in tasks as part of the service.

containerService_loadBalancers :: Lens' ContainerService (Maybe [LoadBalancer])

A list of Elastic Load Balancing load balancer objects, containing the load balancer name, the container name (as it appears in a container definition), and the container port to access from the load balancer.

containerService_tags :: Lens' ContainerService (Maybe [Tag])

The metadata that you apply to the service to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define.

The following basic restrictions apply to tags:

Maximum number of tags per resource - 50

For each resource, each tag key must be unique, and each tag key can have only one value.

Maximum key length - 128 Unicode characters in UTF-8

Maximum value length - 256 Unicode characters in UTF-8

If your tagging schema is used across multiple services and resources, remember that other services may have restrictions on allowed characters. Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following characters: + - = . _ : / @.

Tag keys and values are case-sensitive.

Do not use aws:, AWS:, or any upper or lowercase combination of such as a prefix for either keys or values as it is reserved for Amazon Web Services use. You cannot edit or delete tag keys or values with this prefix. Tags with this prefix do not count against your tags per resource limit.

containerService_healthCheckGracePeriodSeconds :: Lens' ContainerService (Maybe Int)

The period of time, in seconds, that the Amazon ECS service scheduler ignores unhealthy Elastic Load Balancing target health checks after a task has first started.

containerService_serviceRegistries :: Lens' ContainerService (Maybe [ServiceRegistry])

The details of the service discovery registries to assign to this service. For more information, see Service Discovery.

containerService_createdBy :: Lens' ContainerService (Maybe Text)

The principal that created the service.

containerService_schedulingStrategy :: Lens' ContainerService (Maybe SchedulingStrategy)

The scheduling strategy to use for the service. For more information, see Services.

There are two service scheduler strategies available:

REPLICA-The replica scheduling strategy places and maintains the desired number of tasks across your cluster. By default, the service scheduler spreads tasks across Availability Zones. You can use task placement strategies and constraints to customize task placement decisions.

DAEMON-The daemon scheduling strategy deploys exactly one task on each active container instance that meets all of the task placement constraints that you specify in your cluster. The service scheduler also evaluates the task placement constraints for running tasks and will stop tasks that do not meet the placement constraints.

Fargate tasks do not support the DAEMON scheduling strategy.

containerService_taskDefinition :: Lens' ContainerService (Maybe Text)

The task definition to use for tasks in the service. This value is specified when the service is created with CreateService, and it can be modified with UpdateService.

containerService_serviceArn :: Lens' ContainerService (Maybe Text)

The ARN that identifies the service. The ARN contains the arn:aws:ecs namespace, followed by the Region of the service, the Amazon Web Services account ID of the service owner, the service namespace, and then the service name. For example, arn:aws:ecs:region:012345678910:service/my-service.

containerService_propagateTags :: Lens' ContainerService (Maybe PropagateTags)

Specifies whether to propagate the tags from the task definition or the service to the task. If no value is specified, the tags are not propagated.