Network.AWS.ECS.Types.ContainerStateChange

Description

Documentation

data ContainerStateChange

An object representing a change in state for a container.

See: newContainerStateChange smart constructor.

Constructors

ContainerStateChange'

Fields

imageDigest :: Maybe Text

The container image SHA 256 digest.

status :: Maybe Text

The status of the container.

runtimeId :: Maybe Text

The ID of the Docker container.

exitCode :: Maybe Int

The exit code for the container, if the state change is a result of the container exiting.

networkBindings :: Maybe [NetworkBinding]

Any network bindings associated with the container.

reason :: Maybe Text

The reason for the state change.

containerName :: Maybe Text

The name of the container.

newContainerStateChange :: ContainerStateChange

Create a value of ContainerStateChange with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:imageDigest:ContainerStateChange', containerStateChange_imageDigest - The container image SHA 256 digest.

$sel:status:ContainerStateChange', containerStateChange_status - The status of the container.

$sel:runtimeId:ContainerStateChange', containerStateChange_runtimeId - The ID of the Docker container.

$sel:exitCode:ContainerStateChange', containerStateChange_exitCode - The exit code for the container, if the state change is a result of the container exiting.

$sel:networkBindings:ContainerStateChange', containerStateChange_networkBindings - Any network bindings associated with the container.

$sel:reason:ContainerStateChange', containerStateChange_reason - The reason for the state change.

$sel:containerName:ContainerStateChange', containerStateChange_containerName - The name of the container.

containerStateChange_imageDigest :: Lens' ContainerStateChange (Maybe Text)

The container image SHA 256 digest.

containerStateChange_status :: Lens' ContainerStateChange (Maybe Text)

The status of the container.

containerStateChange_runtimeId :: Lens' ContainerStateChange (Maybe Text)

The ID of the Docker container.

containerStateChange_exitCode :: Lens' ContainerStateChange (Maybe Int)

The exit code for the container, if the state change is a result of the container exiting.

containerStateChange_networkBindings :: Lens' ContainerStateChange (Maybe [NetworkBinding])

Any network bindings associated with the container.

containerStateChange_reason :: Lens' ContainerStateChange (Maybe Text)

The reason for the state change.

containerStateChange_containerName :: Lens' ContainerStateChange (Maybe Text)

The name of the container.