Network.AWS.ECS.Types.Deployment

Description

Documentation

data Deployment

The details of an Amazon ECS service deployment. This is used only when a service uses the ECS deployment controller type.

See: newDeployment smart constructor.

Constructors

Deployment'

Fields

rolloutState :: Maybe DeploymentRolloutState

The rolloutState of a service is only returned for services that use the rolling update (ECS) deployment type that are not behind a Classic Load Balancer.

The rollout state of the deployment. When a service deployment is started, it begins in an IN_PROGRESS state. When the service reaches a steady state, the deployment will transition to a COMPLETED state. If the service fails to reach a steady state and circuit breaker is enabled, the deployment will transition to a FAILED state. A deployment in FAILED state will launch no new tasks. For more information, see DeploymentCircuitBreaker.

runningCount :: Maybe Int

The number of tasks in the deployment that are in the RUNNING status.

status :: Maybe Text

The status of the deployment. The following describes each state:

PRIMARY

The most recent deployment of a service.

ACTIVE

A service deployment that still has running tasks, but are in the process of being replaced with a new PRIMARY deployment.

INACTIVE

A deployment that has been completely replaced.

capacityProviderStrategy :: Maybe [CapacityProviderStrategyItem]

The capacity provider strategy that the deployment is using.

networkConfiguration :: Maybe NetworkConfiguration

The VPC subnet and security group configuration for tasks that receive their own elastic network interface by using the awsvpc networking mode.

desiredCount :: Maybe Int

The most recent desired count of tasks that was specified for the service to deploy or maintain.

updatedAt :: Maybe POSIX

The Unix timestamp for when the service deployment was last updated.

launchType :: Maybe LaunchType

The launch type the tasks in the service are using. For more information, see Amazon ECS Launch Types in the Amazon Elastic Container Service Developer Guide.

id :: Maybe Text

The ID of the deployment.

createdAt :: Maybe POSIX

The Unix timestamp for when the service deployment was created.

platformVersion :: Maybe Text

The platform version on which your tasks in the service are running. A platform version is only specified for tasks using the Fargate launch type. If one is not specified, the LATEST platform version is used by default. For more information, see Fargate Platform Versions in the Amazon Elastic Container Service Developer Guide.

pendingCount :: Maybe Int

The number of tasks in the deployment that are in the PENDING status.

taskDefinition :: Maybe Text

The most recent task definition that was specified for the tasks in the service to use.

rolloutStateReason :: Maybe Text

A description of the rollout state of a deployment.

failedTasks :: Maybe Int

The number of consecutively failed tasks in the deployment. A task is considered a failure if the service scheduler can't launch the task, the task doesn't transition to a RUNNING state, or if it fails any of its defined health checks and is stopped.

Once a service deployment has one or more successfully running tasks, the failed task count resets to zero and stops being evaluated.

newDeployment :: Deployment

Create a value of Deployment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:rolloutState:Deployment', deployment_rolloutState - The rolloutState of a service is only returned for services that use the rolling update (ECS) deployment type that are not behind a Classic Load Balancer.

The rollout state of the deployment. When a service deployment is started, it begins in an IN_PROGRESS state. When the service reaches a steady state, the deployment will transition to a COMPLETED state. If the service fails to reach a steady state and circuit breaker is enabled, the deployment will transition to a FAILED state. A deployment in FAILED state will launch no new tasks. For more information, see DeploymentCircuitBreaker.

$sel:runningCount:Deployment', deployment_runningCount - The number of tasks in the deployment that are in the RUNNING status.

$sel:status:Deployment', deployment_status - The status of the deployment. The following describes each state:

PRIMARY

The most recent deployment of a service.

ACTIVE

A service deployment that still has running tasks, but are in the process of being replaced with a new PRIMARY deployment.

INACTIVE

A deployment that has been completely replaced.

$sel:capacityProviderStrategy:Deployment', deployment_capacityProviderStrategy - The capacity provider strategy that the deployment is using.

$sel:networkConfiguration:Deployment', deployment_networkConfiguration - The VPC subnet and security group configuration for tasks that receive their own elastic network interface by using the awsvpc networking mode.

$sel:desiredCount:Deployment', deployment_desiredCount - The most recent desired count of tasks that was specified for the service to deploy or maintain.

$sel:updatedAt:Deployment', deployment_updatedAt - The Unix timestamp for when the service deployment was last updated.

$sel:launchType:Deployment', deployment_launchType - The launch type the tasks in the service are using. For more information, see Amazon ECS Launch Types in the Amazon Elastic Container Service Developer Guide.

$sel:id:Deployment', deployment_id - The ID of the deployment.

$sel:createdAt:Deployment', deployment_createdAt - The Unix timestamp for when the service deployment was created.

$sel:platformVersion:Deployment', deployment_platformVersion - The platform version on which your tasks in the service are running. A platform version is only specified for tasks using the Fargate launch type. If one is not specified, the LATEST platform version is used by default. For more information, see Fargate Platform Versions in the Amazon Elastic Container Service Developer Guide.

$sel:pendingCount:Deployment', deployment_pendingCount - The number of tasks in the deployment that are in the PENDING status.

$sel:taskDefinition:Deployment', deployment_taskDefinition - The most recent task definition that was specified for the tasks in the service to use.

$sel:rolloutStateReason:Deployment', deployment_rolloutStateReason - A description of the rollout state of a deployment.

$sel:failedTasks:Deployment', deployment_failedTasks - The number of consecutively failed tasks in the deployment. A task is considered a failure if the service scheduler can't launch the task, the task doesn't transition to a RUNNING state, or if it fails any of its defined health checks and is stopped.

Once a service deployment has one or more successfully running tasks, the failed task count resets to zero and stops being evaluated.

deployment_rolloutState :: Lens' Deployment (Maybe DeploymentRolloutState)

The rolloutState of a service is only returned for services that use the rolling update (ECS) deployment type that are not behind a Classic Load Balancer.

The rollout state of the deployment. When a service deployment is started, it begins in an IN_PROGRESS state. When the service reaches a steady state, the deployment will transition to a COMPLETED state. If the service fails to reach a steady state and circuit breaker is enabled, the deployment will transition to a FAILED state. A deployment in FAILED state will launch no new tasks. For more information, see DeploymentCircuitBreaker.

deployment_runningCount :: Lens' Deployment (Maybe Int)

The number of tasks in the deployment that are in the RUNNING status.

deployment_status :: Lens' Deployment (Maybe Text)

The status of the deployment. The following describes each state:

PRIMARY

The most recent deployment of a service.

ACTIVE

A service deployment that still has running tasks, but are in the process of being replaced with a new PRIMARY deployment.

INACTIVE

A deployment that has been completely replaced.

deployment_capacityProviderStrategy :: Lens' Deployment (Maybe [CapacityProviderStrategyItem])

The capacity provider strategy that the deployment is using.

deployment_networkConfiguration :: Lens' Deployment (Maybe NetworkConfiguration)

The VPC subnet and security group configuration for tasks that receive their own elastic network interface by using the awsvpc networking mode.

deployment_desiredCount :: Lens' Deployment (Maybe Int)

The most recent desired count of tasks that was specified for the service to deploy or maintain.

deployment_updatedAt :: Lens' Deployment (Maybe UTCTime)

The Unix timestamp for when the service deployment was last updated.

deployment_launchType :: Lens' Deployment (Maybe LaunchType)

The launch type the tasks in the service are using. For more information, see Amazon ECS Launch Types in the Amazon Elastic Container Service Developer Guide.

deployment_id :: Lens' Deployment (Maybe Text)

The ID of the deployment.

deployment_createdAt :: Lens' Deployment (Maybe UTCTime)

The Unix timestamp for when the service deployment was created.

deployment_platformVersion :: Lens' Deployment (Maybe Text)

The platform version on which your tasks in the service are running. A platform version is only specified for tasks using the Fargate launch type. If one is not specified, the LATEST platform version is used by default. For more information, see Fargate Platform Versions in the Amazon Elastic Container Service Developer Guide.

deployment_pendingCount :: Lens' Deployment (Maybe Int)

The number of tasks in the deployment that are in the PENDING status.

deployment_taskDefinition :: Lens' Deployment (Maybe Text)

The most recent task definition that was specified for the tasks in the service to use.

deployment_rolloutStateReason :: Lens' Deployment (Maybe Text)

A description of the rollout state of a deployment.

deployment_failedTasks :: Lens' Deployment (Maybe Int)

The number of consecutively failed tasks in the deployment. A task is considered a failure if the service scheduler can't launch the task, the task doesn't transition to a RUNNING state, or if it fails any of its defined health checks and is stopped.

Once a service deployment has one or more successfully running tasks, the failed task count resets to zero and stops being evaluated.