Network.AWS.ECS.Types.DockerVolumeConfiguration

Description

Documentation

data DockerVolumeConfiguration

This parameter is specified when you are using Docker volumes. Docker volumes are only supported when you are using the EC2 launch type. Windows containers only support the use of the local driver. To use bind mounts, specify a host instead.

See: newDockerVolumeConfiguration smart constructor.

Constructors

DockerVolumeConfiguration'

Fields

labels :: Maybe (HashMap Text Text)

Custom metadata to add to your Docker volume. This parameter maps to Labels in the Create a volume section of the Docker Remote API and the xxlabel option to docker volume create.

scope :: Maybe Scope

The scope for the Docker volume that determines its lifecycle. Docker volumes that are scoped to a task are automatically provisioned when the task starts and destroyed when the task stops. Docker volumes that are scoped as shared persist after the task stops.

driverOpts :: Maybe (HashMap Text Text)

A map of Docker driver-specific options passed through. This parameter maps to DriverOpts in the Create a volume section of the Docker Remote API and the xxopt option to docker volume create.

autoprovision :: Maybe Bool

If this value is true, the Docker volume is created if it does not already exist.

This field is only used if the scope is shared.

driver :: Maybe Text

The Docker volume driver to use. The driver value must match the driver name provided by Docker because it is used for task placement. If the driver was installed using the Docker plugin CLI, use docker plugin ls to retrieve the driver name from your container instance. If the driver was installed using another method, use Docker plugin discovery to retrieve the driver name. For more information, see Docker plugin discovery. This parameter maps to Driver in the Create a volume section of the Docker Remote API and the xxdriver option to docker volume create.

newDockerVolumeConfiguration :: DockerVolumeConfiguration

Create a value of DockerVolumeConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:labels:DockerVolumeConfiguration', dockerVolumeConfiguration_labels - Custom metadata to add to your Docker volume. This parameter maps to Labels in the Create a volume section of the Docker Remote API and the xxlabel option to docker volume create.

$sel:scope:DockerVolumeConfiguration', dockerVolumeConfiguration_scope - The scope for the Docker volume that determines its lifecycle. Docker volumes that are scoped to a task are automatically provisioned when the task starts and destroyed when the task stops. Docker volumes that are scoped as shared persist after the task stops.

$sel:driverOpts:DockerVolumeConfiguration', dockerVolumeConfiguration_driverOpts - A map of Docker driver-specific options passed through. This parameter maps to DriverOpts in the Create a volume section of the Docker Remote API and the xxopt option to docker volume create.

$sel:autoprovision:DockerVolumeConfiguration', dockerVolumeConfiguration_autoprovision - If this value is true, the Docker volume is created if it does not already exist.

This field is only used if the scope is shared.

$sel:driver:DockerVolumeConfiguration', dockerVolumeConfiguration_driver - The Docker volume driver to use. The driver value must match the driver name provided by Docker because it is used for task placement. If the driver was installed using the Docker plugin CLI, use docker plugin ls to retrieve the driver name from your container instance. If the driver was installed using another method, use Docker plugin discovery to retrieve the driver name. For more information, see Docker plugin discovery. This parameter maps to Driver in the Create a volume section of the Docker Remote API and the xxdriver option to docker volume create.

dockerVolumeConfiguration_labels :: Lens' DockerVolumeConfiguration (Maybe (HashMap Text Text))

Custom metadata to add to your Docker volume. This parameter maps to Labels in the Create a volume section of the Docker Remote API and the xxlabel option to docker volume create.

dockerVolumeConfiguration_scope :: Lens' DockerVolumeConfiguration (Maybe Scope)

The scope for the Docker volume that determines its lifecycle. Docker volumes that are scoped to a task are automatically provisioned when the task starts and destroyed when the task stops. Docker volumes that are scoped as shared persist after the task stops.

dockerVolumeConfiguration_driverOpts :: Lens' DockerVolumeConfiguration (Maybe (HashMap Text Text))

A map of Docker driver-specific options passed through. This parameter maps to DriverOpts in the Create a volume section of the Docker Remote API and the xxopt option to docker volume create.

dockerVolumeConfiguration_autoprovision :: Lens' DockerVolumeConfiguration (Maybe Bool)

If this value is true, the Docker volume is created if it does not already exist.

This field is only used if the scope is shared.

dockerVolumeConfiguration_driver :: Lens' DockerVolumeConfiguration (Maybe Text)

The Docker volume driver to use. The driver value must match the driver name provided by Docker because it is used for task placement. If the driver was installed using the Docker plugin CLI, use docker plugin ls to retrieve the driver name from your container instance. If the driver was installed using another method, use Docker plugin discovery to retrieve the driver name. For more information, see Docker plugin discovery. This parameter maps to Driver in the Create a volume section of the Docker Remote API and the xxdriver option to docker volume create.