Network.AWS.ECS.Types.EFSVolumeConfiguration

Description

Documentation

data EFSVolumeConfiguration

This parameter is specified when you are using an Amazon Elastic File System file system for task storage. For more information, see Amazon EFS Volumes in the Amazon Elastic Container Service Developer Guide.

See: newEFSVolumeConfiguration smart constructor.

Constructors

EFSVolumeConfiguration'

Fields

transitEncryptionPort :: Maybe Int

The port to use when sending encrypted data between the Amazon ECS host and the Amazon EFS server. If you do not specify a transit encryption port, it will use the port selection strategy that the Amazon EFS mount helper uses. For more information, see EFS Mount Helper in the Amazon Elastic File System User Guide.

rootDirectory :: Maybe Text

The directory within the Amazon EFS file system to mount as the root directory inside the host. If this parameter is omitted, the root of the Amazon EFS volume will be used. Specifying / will have the same effect as omitting this parameter.

If an EFS access point is specified in the authorizationConfig, the root directory parameter must either be omitted or set to / which will enforce the path set on the EFS access point.

authorizationConfig :: Maybe EFSAuthorizationConfig

The authorization configuration details for the Amazon EFS file system.

transitEncryption :: Maybe EFSTransitEncryption

Whether or not to enable encryption for Amazon EFS data in transit between the Amazon ECS host and the Amazon EFS server. Transit encryption must be enabled if Amazon EFS IAM authorization is used. If this parameter is omitted, the default value of DISABLED is used. For more information, see Encrypting Data in Transit in the Amazon Elastic File System User Guide.

fileSystemId :: Text

The Amazon EFS file system ID to use.

newEFSVolumeConfiguration

Arguments

:: Text
$sel:fileSystemId:EFSVolumeConfiguration'

-> EFSVolumeConfiguration

Create a value of EFSVolumeConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:transitEncryptionPort:EFSVolumeConfiguration', eFSVolumeConfiguration_transitEncryptionPort - The port to use when sending encrypted data between the Amazon ECS host and the Amazon EFS server. If you do not specify a transit encryption port, it will use the port selection strategy that the Amazon EFS mount helper uses. For more information, see EFS Mount Helper in the Amazon Elastic File System User Guide.

$sel:rootDirectory:EFSVolumeConfiguration', eFSVolumeConfiguration_rootDirectory - The directory within the Amazon EFS file system to mount as the root directory inside the host. If this parameter is omitted, the root of the Amazon EFS volume will be used. Specifying / will have the same effect as omitting this parameter.

If an EFS access point is specified in the authorizationConfig, the root directory parameter must either be omitted or set to / which will enforce the path set on the EFS access point.

$sel:authorizationConfig:EFSVolumeConfiguration', eFSVolumeConfiguration_authorizationConfig - The authorization configuration details for the Amazon EFS file system.

$sel:transitEncryption:EFSVolumeConfiguration', eFSVolumeConfiguration_transitEncryption - Whether or not to enable encryption for Amazon EFS data in transit between the Amazon ECS host and the Amazon EFS server. Transit encryption must be enabled if Amazon EFS IAM authorization is used. If this parameter is omitted, the default value of DISABLED is used. For more information, see Encrypting Data in Transit in the Amazon Elastic File System User Guide.

$sel:fileSystemId:EFSVolumeConfiguration', eFSVolumeConfiguration_fileSystemId - The Amazon EFS file system ID to use.

eFSVolumeConfiguration_transitEncryptionPort :: Lens' EFSVolumeConfiguration (Maybe Int)

The port to use when sending encrypted data between the Amazon ECS host and the Amazon EFS server. If you do not specify a transit encryption port, it will use the port selection strategy that the Amazon EFS mount helper uses. For more information, see EFS Mount Helper in the Amazon Elastic File System User Guide.

eFSVolumeConfiguration_rootDirectory :: Lens' EFSVolumeConfiguration (Maybe Text)

The directory within the Amazon EFS file system to mount as the root directory inside the host. If this parameter is omitted, the root of the Amazon EFS volume will be used. Specifying / will have the same effect as omitting this parameter.

If an EFS access point is specified in the authorizationConfig, the root directory parameter must either be omitted or set to / which will enforce the path set on the EFS access point.

eFSVolumeConfiguration_authorizationConfig :: Lens' EFSVolumeConfiguration (Maybe EFSAuthorizationConfig)

The authorization configuration details for the Amazon EFS file system.

eFSVolumeConfiguration_transitEncryption :: Lens' EFSVolumeConfiguration (Maybe EFSTransitEncryption)

Whether or not to enable encryption for Amazon EFS data in transit between the Amazon ECS host and the Amazon EFS server. Transit encryption must be enabled if Amazon EFS IAM authorization is used. If this parameter is omitted, the default value of DISABLED is used. For more information, see Encrypting Data in Transit in the Amazon Elastic File System User Guide.

eFSVolumeConfiguration_fileSystemId :: Lens' EFSVolumeConfiguration Text

The Amazon EFS file system ID to use.