Network.AWS.ECS.Types.ExecuteCommandConfiguration

Description

Documentation

data ExecuteCommandConfiguration

The details of the execute command configuration.

See: newExecuteCommandConfiguration smart constructor.

Constructors

ExecuteCommandConfiguration'

Fields

logging :: Maybe ExecuteCommandLogging

The log setting to use for redirecting logs for your execute command results. The following log settings are available.

NONE: The execute command session is not logged.

DEFAULT: The awslogs configuration in the task definition is used. If no logging parameter is specified, it defaults to this value. If no awslogs log driver is configured in the task definition, the output won't be logged.

OVERRIDE: Specify the logging details as a part of logConfiguration. If the OVERRIDE logging option is specified, the logConfiguration is required.

kmsKeyId :: Maybe Text

Specify an Key Management Service key ID to encrypt the data between the local client and the container.

logConfiguration :: Maybe ExecuteCommandLogConfiguration

The log configuration for the results of the execute command actions. The logs can be sent to CloudWatch Logs or an Amazon S3 bucket. When logging=OVERRIDE is specified, a logConfiguration must be provided.

newExecuteCommandConfiguration :: ExecuteCommandConfiguration

Create a value of ExecuteCommandConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:logging:ExecuteCommandConfiguration', executeCommandConfiguration_logging - The log setting to use for redirecting logs for your execute command results. The following log settings are available.

NONE: The execute command session is not logged.

DEFAULT: The awslogs configuration in the task definition is used. If no logging parameter is specified, it defaults to this value. If no awslogs log driver is configured in the task definition, the output won't be logged.

OVERRIDE: Specify the logging details as a part of logConfiguration. If the OVERRIDE logging option is specified, the logConfiguration is required.

$sel:kmsKeyId:ExecuteCommandConfiguration', executeCommandConfiguration_kmsKeyId - Specify an Key Management Service key ID to encrypt the data between the local client and the container.

$sel:logConfiguration:ExecuteCommandConfiguration', executeCommandConfiguration_logConfiguration - The log configuration for the results of the execute command actions. The logs can be sent to CloudWatch Logs or an Amazon S3 bucket. When logging=OVERRIDE is specified, a logConfiguration must be provided.

executeCommandConfiguration_logging :: Lens' ExecuteCommandConfiguration (Maybe ExecuteCommandLogging)

The log setting to use for redirecting logs for your execute command results. The following log settings are available.

NONE: The execute command session is not logged.

DEFAULT: The awslogs configuration in the task definition is used. If no logging parameter is specified, it defaults to this value. If no awslogs log driver is configured in the task definition, the output won't be logged.

OVERRIDE: Specify the logging details as a part of logConfiguration. If the OVERRIDE logging option is specified, the logConfiguration is required.

executeCommandConfiguration_kmsKeyId :: Lens' ExecuteCommandConfiguration (Maybe Text)

Specify an Key Management Service key ID to encrypt the data between the local client and the container.

executeCommandConfiguration_logConfiguration :: Lens' ExecuteCommandConfiguration (Maybe ExecuteCommandLogConfiguration)

The log configuration for the results of the execute command actions. The logs can be sent to CloudWatch Logs or an Amazon S3 bucket. When logging=OVERRIDE is specified, a logConfiguration must be provided.