Network.AWS.ECS.Types.ManagedAgentStateChange

Description

Documentation

data ManagedAgentStateChange

An object representing a change in state for a managed agent.

See: newManagedAgentStateChange smart constructor.

Constructors

ManagedAgentStateChange'

Fields

reason :: Maybe Text

The reason for the status of the managed agent.

containerName :: Text

The name of the container associated with the managed agent.

managedAgentName :: ManagedAgentName

The name of the managed agent.

status :: Text

The status of the managed agent.

newManagedAgentStateChange

Arguments

:: Text
$sel:containerName:ManagedAgentStateChange'

-> ManagedAgentName
$sel:managedAgentName:ManagedAgentStateChange'

-> Text
$sel:status:ManagedAgentStateChange'

-> ManagedAgentStateChange

Create a value of ManagedAgentStateChange with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reason:ManagedAgentStateChange', managedAgentStateChange_reason - The reason for the status of the managed agent.

$sel:containerName:ManagedAgentStateChange', managedAgentStateChange_containerName - The name of the container associated with the managed agent.

$sel:managedAgentName:ManagedAgentStateChange', managedAgentStateChange_managedAgentName - The name of the managed agent.

$sel:status:ManagedAgentStateChange', managedAgentStateChange_status - The status of the managed agent.

managedAgentStateChange_reason :: Lens' ManagedAgentStateChange (Maybe Text)

The reason for the status of the managed agent.

managedAgentStateChange_containerName :: Lens' ManagedAgentStateChange Text

The name of the container associated with the managed agent.

managedAgentStateChange_managedAgentName :: Lens' ManagedAgentStateChange ManagedAgentName

The name of the managed agent.

managedAgentStateChange_status :: Lens' ManagedAgentStateChange Text

The status of the managed agent.