Network.AWS.ECS.Types.MountPoint

Description

Documentation

data MountPoint

Details on a volume mount point that is used in a container definition.

See: newMountPoint smart constructor.

Constructors

MountPoint'

Fields

readOnly :: Maybe Bool

If this value is true, the container has read-only access to the volume. If this value is false, then the container can write to the volume. The default value is false.

sourceVolume :: Maybe Text

The name of the volume to mount. Must be a volume name referenced in the name parameter of task definition volume.

containerPath :: Maybe Text

The path on the container to mount the host volume at.

newMountPoint :: MountPoint

Create a value of MountPoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:readOnly:MountPoint', mountPoint_readOnly - If this value is true, the container has read-only access to the volume. If this value is false, then the container can write to the volume. The default value is false.

$sel:sourceVolume:MountPoint', mountPoint_sourceVolume - The name of the volume to mount. Must be a volume name referenced in the name parameter of task definition volume.

$sel:containerPath:MountPoint', mountPoint_containerPath - The path on the container to mount the host volume at.

mountPoint_readOnly :: Lens' MountPoint (Maybe Bool)

If this value is true, the container has read-only access to the volume. If this value is false, then the container can write to the volume. The default value is false.

mountPoint_sourceVolume :: Lens' MountPoint (Maybe Text)

The name of the volume to mount. Must be a volume name referenced in the name parameter of task definition volume.

mountPoint_containerPath :: Lens' MountPoint (Maybe Text)

The path on the container to mount the host volume at.