Network.AWS.ECS.Types.ProxyConfiguration

Description

Documentation

data ProxyConfiguration

The configuration details for the App Mesh proxy.

For tasks using the EC2 launch type, the container instances require at least version 1.26.0 of the container agent and at least version 1.26.0-1 of the ecs-init package to enable a proxy configuration. If your container instances are launched from the Amazon ECS-optimized AMI version 20190301 or later, then they contain the required versions of the container agent and ecs-init. For more information, see Amazon ECS-optimized Linux AMI

See: newProxyConfiguration smart constructor.

Constructors

ProxyConfiguration'

Fields

properties :: Maybe [KeyValuePair]

The set of network configuration parameters to provide the Container Network Interface (CNI) plugin, specified as key-value pairs.

IgnoredUID - (Required) The user ID (UID) of the proxy container as defined by the user parameter in a container definition. This is used to ensure the proxy ignores its own traffic. If IgnoredGID is specified, this field can be empty.

IgnoredGID - (Required) The group ID (GID) of the proxy container as defined by the user parameter in a container definition. This is used to ensure the proxy ignores its own traffic. If IgnoredUID is specified, this field can be empty.

AppPorts - (Required) The list of ports that the application uses. Network traffic to these ports is forwarded to the ProxyIngressPort and ProxyEgressPort.

ProxyIngressPort - (Required) Specifies the port that incoming traffic to the AppPorts is directed to.

ProxyEgressPort - (Required) Specifies the port that outgoing traffic from the AppPorts is directed to.

EgressIgnoredPorts - (Required) The egress traffic going to the specified ports is ignored and not redirected to the ProxyEgressPort. It can be an empty list.

EgressIgnoredIPs - (Required) The egress traffic going to the specified IP addresses is ignored and not redirected to the ProxyEgressPort. It can be an empty list.

type' :: Maybe ProxyConfigurationType

The proxy type. The only supported value is APPMESH.

containerName :: Text

The name of the container that will serve as the App Mesh proxy.

newProxyConfiguration

Arguments

:: Text
$sel:containerName:ProxyConfiguration'

-> ProxyConfiguration

Create a value of ProxyConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:properties:ProxyConfiguration', proxyConfiguration_properties - The set of network configuration parameters to provide the Container Network Interface (CNI) plugin, specified as key-value pairs.

IgnoredUID - (Required) The user ID (UID) of the proxy container as defined by the user parameter in a container definition. This is used to ensure the proxy ignores its own traffic. If IgnoredGID is specified, this field can be empty.

IgnoredGID - (Required) The group ID (GID) of the proxy container as defined by the user parameter in a container definition. This is used to ensure the proxy ignores its own traffic. If IgnoredUID is specified, this field can be empty.

AppPorts - (Required) The list of ports that the application uses. Network traffic to these ports is forwarded to the ProxyIngressPort and ProxyEgressPort.

ProxyIngressPort - (Required) Specifies the port that incoming traffic to the AppPorts is directed to.

ProxyEgressPort - (Required) Specifies the port that outgoing traffic from the AppPorts is directed to.

EgressIgnoredPorts - (Required) The egress traffic going to the specified ports is ignored and not redirected to the ProxyEgressPort. It can be an empty list.

EgressIgnoredIPs - (Required) The egress traffic going to the specified IP addresses is ignored and not redirected to the ProxyEgressPort. It can be an empty list.

$sel:type':ProxyConfiguration', proxyConfiguration_type - The proxy type. The only supported value is APPMESH.

$sel:containerName:ProxyConfiguration', proxyConfiguration_containerName - The name of the container that will serve as the App Mesh proxy.

proxyConfiguration_properties :: Lens' ProxyConfiguration (Maybe [KeyValuePair])

The set of network configuration parameters to provide the Container Network Interface (CNI) plugin, specified as key-value pairs.

IgnoredUID - (Required) The user ID (UID) of the proxy container as defined by the user parameter in a container definition. This is used to ensure the proxy ignores its own traffic. If IgnoredGID is specified, this field can be empty.

IgnoredGID - (Required) The group ID (GID) of the proxy container as defined by the user parameter in a container definition. This is used to ensure the proxy ignores its own traffic. If IgnoredUID is specified, this field can be empty.

AppPorts - (Required) The list of ports that the application uses. Network traffic to these ports is forwarded to the ProxyIngressPort and ProxyEgressPort.

ProxyIngressPort - (Required) Specifies the port that incoming traffic to the AppPorts is directed to.

ProxyEgressPort - (Required) Specifies the port that outgoing traffic from the AppPorts is directed to.

EgressIgnoredPorts - (Required) The egress traffic going to the specified ports is ignored and not redirected to the ProxyEgressPort. It can be an empty list.

EgressIgnoredIPs - (Required) The egress traffic going to the specified IP addresses is ignored and not redirected to the ProxyEgressPort. It can be an empty list.

proxyConfiguration_type :: Lens' ProxyConfiguration (Maybe ProxyConfigurationType)

The proxy type. The only supported value is APPMESH.

proxyConfiguration_containerName :: Lens' ProxyConfiguration Text

The name of the container that will serve as the App Mesh proxy.