Network.AWS.ECS.Types.Setting

Description

Documentation

data Setting

The current account setting for a resource.

See: newSetting smart constructor.

Constructors

Setting'

Fields

value :: Maybe Text

Whether the account setting is enabled or disabled for the specified resource.

name :: Maybe SettingName

The Amazon ECS resource name.

principalArn :: Maybe Text

The ARN of the principal, which can be an IAM user, IAM role, or the root user. If this field is omitted, the authenticated user is assumed.

newSetting :: Setting

Create a value of Setting with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:Setting', setting_value - Whether the account setting is enabled or disabled for the specified resource.

$sel:name:Setting', setting_name - The Amazon ECS resource name.

$sel:principalArn:Setting', setting_principalArn - The ARN of the principal, which can be an IAM user, IAM role, or the root user. If this field is omitted, the authenticated user is assumed.

setting_value :: Lens' Setting (Maybe Text)

Whether the account setting is enabled or disabled for the specified resource.

setting_name :: Lens' Setting (Maybe SettingName)

The Amazon ECS resource name.

setting_principalArn :: Lens' Setting (Maybe Text)

The ARN of the principal, which can be an IAM user, IAM role, or the root user. If this field is omitted, the authenticated user is assumed.