Network.AWS.ECS.Types.TaskOverride

Description

Documentation

data TaskOverride

The overrides associated with a task.

See: newTaskOverride smart constructor.

Constructors

TaskOverride'

Fields

taskRoleArn :: Maybe Text

The Amazon Resource Name (ARN) of the IAM role that containers in this task can assume. All containers in this task are granted the permissions that are specified in this role.

memory :: Maybe Text

The memory override for the task.

inferenceAcceleratorOverrides :: Maybe [InferenceAcceleratorOverride]

The Elastic Inference accelerator override for the task.

executionRoleArn :: Maybe Text

The Amazon Resource Name (ARN) of the task execution IAM role override for the task.

containerOverrides :: Maybe [ContainerOverride]

One or more container overrides sent to a task.

ephemeralStorage :: Maybe EphemeralStorage

The ephemeral storage setting override for the task.

This parameter is only supported for tasks hosted on Fargate using platform version 1.4.0 or later.

cpu :: Maybe Text

The cpu override for the task.

newTaskOverride :: TaskOverride

Create a value of TaskOverride with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:taskRoleArn:TaskOverride', taskOverride_taskRoleArn - The Amazon Resource Name (ARN) of the IAM role that containers in this task can assume. All containers in this task are granted the permissions that are specified in this role.

$sel:memory:TaskOverride', taskOverride_memory - The memory override for the task.

$sel:inferenceAcceleratorOverrides:TaskOverride', taskOverride_inferenceAcceleratorOverrides - The Elastic Inference accelerator override for the task.

$sel:executionRoleArn:TaskOverride', taskOverride_executionRoleArn - The Amazon Resource Name (ARN) of the task execution IAM role override for the task.

$sel:containerOverrides:TaskOverride', taskOverride_containerOverrides - One or more container overrides sent to a task.

$sel:ephemeralStorage:TaskOverride', taskOverride_ephemeralStorage - The ephemeral storage setting override for the task.

This parameter is only supported for tasks hosted on Fargate using platform version 1.4.0 or later.

$sel:cpu:TaskOverride', taskOverride_cpu - The cpu override for the task.

taskOverride_taskRoleArn :: Lens' TaskOverride (Maybe Text)

The Amazon Resource Name (ARN) of the IAM role that containers in this task can assume. All containers in this task are granted the permissions that are specified in this role.

taskOverride_memory :: Lens' TaskOverride (Maybe Text)

The memory override for the task.

taskOverride_inferenceAcceleratorOverrides :: Lens' TaskOverride (Maybe [InferenceAcceleratorOverride])

The Elastic Inference accelerator override for the task.

taskOverride_executionRoleArn :: Lens' TaskOverride (Maybe Text)

The Amazon Resource Name (ARN) of the task execution IAM role override for the task.

taskOverride_containerOverrides :: Lens' TaskOverride (Maybe [ContainerOverride])

One or more container overrides sent to a task.

taskOverride_ephemeralStorage :: Lens' TaskOverride (Maybe EphemeralStorage)

The ephemeral storage setting override for the task.

This parameter is only supported for tasks hosted on Fargate using platform version 1.4.0 or later.

taskOverride_cpu :: Lens' TaskOverride (Maybe Text)

The cpu override for the task.