Network.AWS.ECS.Types.TaskSet

Description

Documentation

data TaskSet

Information about a set of Amazon ECS tasks in either an CodeDeploy or an EXTERNAL deployment. An Amazon ECS task set includes details such as the desired number of tasks, how many tasks are running, and whether the task set serves production traffic.

See: newTaskSet smart constructor.

Constructors

TaskSet'

Fields

clusterArn :: Maybe Text

The Amazon Resource Name (ARN) of the cluster that the service that hosts the task set exists in.

stabilityStatusAt :: Maybe POSIX

The Unix timestamp for when the task set stability status was retrieved.

runningCount :: Maybe Int

The number of tasks in the task set that are in the RUNNING status during a deployment. A task in the RUNNING state is running and ready for use.

status :: Maybe Text

The status of the task set. The following describes each state:

PRIMARY

The task set is serving production traffic.

ACTIVE

The task set is not serving production traffic.

DRAINING

The tasks in the task set are being stopped and their corresponding targets are being deregistered from their target group.

stabilityStatus :: Maybe StabilityStatus

The stability status, which indicates whether the task set has reached a steady state. If the following conditions are met, the task set will be in STEADY_STATE:

The task runningCount is equal to the computedDesiredCount.

The pendingCount is 0.

There are no tasks running on container instances in the DRAINING status.

All tasks are reporting a healthy status from the load balancers, service discovery, and container health checks.

If any of those conditions are not met, the stability status returns STABILIZING.

capacityProviderStrategy :: Maybe [CapacityProviderStrategyItem]

The capacity provider strategy associated with the task set.

networkConfiguration :: Maybe NetworkConfiguration

The network configuration for the task set.

updatedAt :: Maybe POSIX

The Unix timestamp for when the task set was last updated.

launchType :: Maybe LaunchType

The launch type the tasks in the task set are using. For more information, see Amazon ECS launch types in the Amazon Elastic Container Service Developer Guide.

id :: Maybe Text

The ID of the task set.

createdAt :: Maybe POSIX

The Unix timestamp for when the task set was created.

platformVersion :: Maybe Text

The Fargate platform version on which the tasks in the task set are running. A platform version is only specified for tasks run on Fargate. For more information, see Fargate platform versions in the Amazon Elastic Container Service Developer Guide.

startedBy :: Maybe Text

The tag specified when a task set is started. If the task set is created by an CodeDeploy deployment, the startedBy parameter is CODE_DEPLOY. For a task set created for an external deployment, the startedBy field isn't used.

computedDesiredCount :: Maybe Int

The computed desired count for the task set. This is calculated by multiplying the service's desiredCount by the task set's scale percentage. The result is always rounded up. For example, if the computed desired count is 1.2, it rounds up to 2 tasks.

pendingCount :: Maybe Int

The number of tasks in the task set that are in the PENDING status during a deployment. A task in the PENDING state is preparing to enter the RUNNING state. A task set enters the PENDING status when it launches for the first time or when it is restarted after being in the STOPPED state.

loadBalancers :: Maybe [LoadBalancer]

Details on a load balancer that is used with a task set.

tags :: Maybe [Tag]

The metadata that you apply to the task set to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define.

The following basic restrictions apply to tags:

Maximum number of tags per resource - 50

For each resource, each tag key must be unique, and each tag key can have only one value.

Maximum key length - 128 Unicode characters in UTF-8

Maximum value length - 256 Unicode characters in UTF-8

If your tagging schema is used across multiple services and resources, remember that other services may have restrictions on allowed characters. Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following characters: + - = . _ : / @.

Tag keys and values are case-sensitive.

Do not use aws:, AWS:, or any upper or lowercase combination of such as a prefix for either keys or values as it is reserved for Amazon Web Services use. You cannot edit or delete tag keys or values with this prefix. Tags with this prefix do not count against your tags per resource limit.

serviceRegistries :: Maybe [ServiceRegistry]

The details of the service discovery registries to assign to this task set. For more information, see Service discovery.

scale :: Maybe Scale

A floating-point percentage of the desired number of tasks to place and keep running in the task set.

taskDefinition :: Maybe Text

The task definition the task set is using.

serviceArn :: Maybe Text

The Amazon Resource Name (ARN) of the service the task set exists in.

externalId :: Maybe Text

The external ID associated with the task set.

If a task set is created by an CodeDeploy deployment, the externalId parameter contains the CodeDeploy deployment ID.

If a task set is created for an external deployment and is associated with a service discovery registry, the externalId parameter contains the ECS_TASK_SET_EXTERNAL_ID Cloud Map attribute.

taskSetArn :: Maybe Text

The Amazon Resource Name (ARN) of the task set.

newTaskSet :: TaskSet

Create a value of TaskSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clusterArn:TaskSet', taskSet_clusterArn - The Amazon Resource Name (ARN) of the cluster that the service that hosts the task set exists in.

$sel:stabilityStatusAt:TaskSet', taskSet_stabilityStatusAt - The Unix timestamp for when the task set stability status was retrieved.

$sel:runningCount:TaskSet', taskSet_runningCount - The number of tasks in the task set that are in the RUNNING status during a deployment. A task in the RUNNING state is running and ready for use.

$sel:status:TaskSet', taskSet_status - The status of the task set. The following describes each state:

PRIMARY

The task set is serving production traffic.

ACTIVE

The task set is not serving production traffic.

DRAINING

The tasks in the task set are being stopped and their corresponding targets are being deregistered from their target group.

$sel:stabilityStatus:TaskSet', taskSet_stabilityStatus - The stability status, which indicates whether the task set has reached a steady state. If the following conditions are met, the task set will be in STEADY_STATE:

The task runningCount is equal to the computedDesiredCount.

The pendingCount is 0.

There are no tasks running on container instances in the DRAINING status.

All tasks are reporting a healthy status from the load balancers, service discovery, and container health checks.

If any of those conditions are not met, the stability status returns STABILIZING.

$sel:capacityProviderStrategy:TaskSet', taskSet_capacityProviderStrategy - The capacity provider strategy associated with the task set.

$sel:networkConfiguration:TaskSet', taskSet_networkConfiguration - The network configuration for the task set.

$sel:updatedAt:TaskSet', taskSet_updatedAt - The Unix timestamp for when the task set was last updated.

$sel:launchType:TaskSet', taskSet_launchType - The launch type the tasks in the task set are using. For more information, see Amazon ECS launch types in the Amazon Elastic Container Service Developer Guide.

$sel:id:TaskSet', taskSet_id - The ID of the task set.

$sel:createdAt:TaskSet', taskSet_createdAt - The Unix timestamp for when the task set was created.

$sel:platformVersion:TaskSet', taskSet_platformVersion - The Fargate platform version on which the tasks in the task set are running. A platform version is only specified for tasks run on Fargate. For more information, see Fargate platform versions in the Amazon Elastic Container Service Developer Guide.

$sel:startedBy:TaskSet', taskSet_startedBy - The tag specified when a task set is started. If the task set is created by an CodeDeploy deployment, the startedBy parameter is CODE_DEPLOY. For a task set created for an external deployment, the startedBy field isn't used.

$sel:computedDesiredCount:TaskSet', taskSet_computedDesiredCount - The computed desired count for the task set. This is calculated by multiplying the service's desiredCount by the task set's scale percentage. The result is always rounded up. For example, if the computed desired count is 1.2, it rounds up to 2 tasks.

$sel:pendingCount:TaskSet', taskSet_pendingCount - The number of tasks in the task set that are in the PENDING status during a deployment. A task in the PENDING state is preparing to enter the RUNNING state. A task set enters the PENDING status when it launches for the first time or when it is restarted after being in the STOPPED state.

$sel:loadBalancers:TaskSet', taskSet_loadBalancers - Details on a load balancer that is used with a task set.

$sel:tags:TaskSet', taskSet_tags - The metadata that you apply to the task set to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define.

The following basic restrictions apply to tags:

Maximum number of tags per resource - 50

For each resource, each tag key must be unique, and each tag key can have only one value.

Maximum key length - 128 Unicode characters in UTF-8

Maximum value length - 256 Unicode characters in UTF-8

If your tagging schema is used across multiple services and resources, remember that other services may have restrictions on allowed characters. Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following characters: + - = . _ : / @.

Tag keys and values are case-sensitive.

Do not use aws:, AWS:, or any upper or lowercase combination of such as a prefix for either keys or values as it is reserved for Amazon Web Services use. You cannot edit or delete tag keys or values with this prefix. Tags with this prefix do not count against your tags per resource limit.

$sel:serviceRegistries:TaskSet', taskSet_serviceRegistries - The details of the service discovery registries to assign to this task set. For more information, see Service discovery.

$sel:scale:TaskSet', taskSet_scale - A floating-point percentage of the desired number of tasks to place and keep running in the task set.

$sel:taskDefinition:TaskSet', taskSet_taskDefinition - The task definition the task set is using.

$sel:serviceArn:TaskSet', taskSet_serviceArn - The Amazon Resource Name (ARN) of the service the task set exists in.

$sel:externalId:TaskSet', taskSet_externalId - The external ID associated with the task set.

If a task set is created by an CodeDeploy deployment, the externalId parameter contains the CodeDeploy deployment ID.

If a task set is created for an external deployment and is associated with a service discovery registry, the externalId parameter contains the ECS_TASK_SET_EXTERNAL_ID Cloud Map attribute.

$sel:taskSetArn:TaskSet', taskSet_taskSetArn - The Amazon Resource Name (ARN) of the task set.

taskSet_clusterArn :: Lens' TaskSet (Maybe Text)

The Amazon Resource Name (ARN) of the cluster that the service that hosts the task set exists in.

taskSet_stabilityStatusAt :: Lens' TaskSet (Maybe UTCTime)

The Unix timestamp for when the task set stability status was retrieved.

taskSet_runningCount :: Lens' TaskSet (Maybe Int)

The number of tasks in the task set that are in the RUNNING status during a deployment. A task in the RUNNING state is running and ready for use.

taskSet_status :: Lens' TaskSet (Maybe Text)

The status of the task set. The following describes each state:

PRIMARY

The task set is serving production traffic.

ACTIVE

The task set is not serving production traffic.

DRAINING

The tasks in the task set are being stopped and their corresponding targets are being deregistered from their target group.

taskSet_stabilityStatus :: Lens' TaskSet (Maybe StabilityStatus)

The stability status, which indicates whether the task set has reached a steady state. If the following conditions are met, the task set will be in STEADY_STATE:

The task runningCount is equal to the computedDesiredCount.

The pendingCount is 0.

There are no tasks running on container instances in the DRAINING status.

All tasks are reporting a healthy status from the load balancers, service discovery, and container health checks.

If any of those conditions are not met, the stability status returns STABILIZING.

taskSet_capacityProviderStrategy :: Lens' TaskSet (Maybe [CapacityProviderStrategyItem])

The capacity provider strategy associated with the task set.

taskSet_networkConfiguration :: Lens' TaskSet (Maybe NetworkConfiguration)

The network configuration for the task set.

taskSet_updatedAt :: Lens' TaskSet (Maybe UTCTime)

The Unix timestamp for when the task set was last updated.

taskSet_launchType :: Lens' TaskSet (Maybe LaunchType)

The launch type the tasks in the task set are using. For more information, see Amazon ECS launch types in the Amazon Elastic Container Service Developer Guide.

taskSet_id :: Lens' TaskSet (Maybe Text)

The ID of the task set.

taskSet_createdAt :: Lens' TaskSet (Maybe UTCTime)

The Unix timestamp for when the task set was created.

taskSet_platformVersion :: Lens' TaskSet (Maybe Text)

The Fargate platform version on which the tasks in the task set are running. A platform version is only specified for tasks run on Fargate. For more information, see Fargate platform versions in the Amazon Elastic Container Service Developer Guide.

taskSet_startedBy :: Lens' TaskSet (Maybe Text)

The tag specified when a task set is started. If the task set is created by an CodeDeploy deployment, the startedBy parameter is CODE_DEPLOY. For a task set created for an external deployment, the startedBy field isn't used.

taskSet_computedDesiredCount :: Lens' TaskSet (Maybe Int)

The computed desired count for the task set. This is calculated by multiplying the service's desiredCount by the task set's scale percentage. The result is always rounded up. For example, if the computed desired count is 1.2, it rounds up to 2 tasks.

taskSet_pendingCount :: Lens' TaskSet (Maybe Int)

The number of tasks in the task set that are in the PENDING status during a deployment. A task in the PENDING state is preparing to enter the RUNNING state. A task set enters the PENDING status when it launches for the first time or when it is restarted after being in the STOPPED state.

taskSet_loadBalancers :: Lens' TaskSet (Maybe [LoadBalancer])

Details on a load balancer that is used with a task set.

taskSet_tags :: Lens' TaskSet (Maybe [Tag])

The metadata that you apply to the task set to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define.

The following basic restrictions apply to tags:

Maximum number of tags per resource - 50

For each resource, each tag key must be unique, and each tag key can have only one value.

Maximum key length - 128 Unicode characters in UTF-8

Maximum value length - 256 Unicode characters in UTF-8

If your tagging schema is used across multiple services and resources, remember that other services may have restrictions on allowed characters. Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following characters: + - = . _ : / @.

Tag keys and values are case-sensitive.

Do not use aws:, AWS:, or any upper or lowercase combination of such as a prefix for either keys or values as it is reserved for Amazon Web Services use. You cannot edit or delete tag keys or values with this prefix. Tags with this prefix do not count against your tags per resource limit.

taskSet_serviceRegistries :: Lens' TaskSet (Maybe [ServiceRegistry])

The details of the service discovery registries to assign to this task set. For more information, see Service discovery.

taskSet_scale :: Lens' TaskSet (Maybe Scale)

A floating-point percentage of the desired number of tasks to place and keep running in the task set.

taskSet_taskDefinition :: Lens' TaskSet (Maybe Text)

The task definition the task set is using.

taskSet_serviceArn :: Lens' TaskSet (Maybe Text)

The Amazon Resource Name (ARN) of the service the task set exists in.

taskSet_externalId :: Lens' TaskSet (Maybe Text)

The external ID associated with the task set.

If a task set is created by an CodeDeploy deployment, the externalId parameter contains the CodeDeploy deployment ID.

If a task set is created for an external deployment and is associated with a service discovery registry, the externalId parameter contains the ECS_TASK_SET_EXTERNAL_ID Cloud Map attribute.

taskSet_taskSetArn :: Lens' TaskSet (Maybe Text)

The Amazon Resource Name (ARN) of the task set.