Network.AWS.ECS.Types.VolumeFrom

Description

Documentation

data VolumeFrom

Details on a data volume from another container in the same task definition.

See: newVolumeFrom smart constructor.

Constructors

VolumeFrom'

Fields

readOnly :: Maybe Bool

If this value is true, the container has read-only access to the volume. If this value is false, then the container can write to the volume. The default value is false.

sourceContainer :: Maybe Text

The name of another container within the same task definition from which to mount volumes.

newVolumeFrom :: VolumeFrom

Create a value of VolumeFrom with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:readOnly:VolumeFrom', volumeFrom_readOnly - If this value is true, the container has read-only access to the volume. If this value is false, then the container can write to the volume. The default value is false.

$sel:sourceContainer:VolumeFrom', volumeFrom_sourceContainer - The name of another container within the same task definition from which to mount volumes.

volumeFrom_readOnly :: Lens' VolumeFrom (Maybe Bool)

If this value is true, the container has read-only access to the volume. If this value is false, then the container can write to the volume. The default value is false.

volumeFrom_sourceContainer :: Lens' VolumeFrom (Maybe Text)

The name of another container within the same task definition from which to mount volumes.