Network.AWS.ECS.UpdateCapacityProvider

Description

Modifies the parameters for a capacity provider.

Creating a Request

data UpdateCapacityProvider

See: newUpdateCapacityProvider smart constructor.

Constructors

UpdateCapacityProvider'

Fields

name :: Text

The name of the capacity provider to update.

autoScalingGroupProvider :: AutoScalingGroupProviderUpdate

An object representing the parameters to update for the Auto Scaling group capacity provider.

newUpdateCapacityProvider

Arguments

:: Text
$sel:name:UpdateCapacityProvider'

-> AutoScalingGroupProviderUpdate
$sel:autoScalingGroupProvider:UpdateCapacityProvider'

-> UpdateCapacityProvider

Create a value of UpdateCapacityProvider with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:UpdateCapacityProvider', updateCapacityProvider_name - The name of the capacity provider to update.

$sel:autoScalingGroupProvider:UpdateCapacityProvider', updateCapacityProvider_autoScalingGroupProvider - An object representing the parameters to update for the Auto Scaling group capacity provider.

Request Lenses

updateCapacityProvider_name :: Lens' UpdateCapacityProvider Text

The name of the capacity provider to update.

updateCapacityProvider_autoScalingGroupProvider :: Lens' UpdateCapacityProvider AutoScalingGroupProviderUpdate

An object representing the parameters to update for the Auto Scaling group capacity provider.

Destructuring the Response

data UpdateCapacityProviderResponse

See: newUpdateCapacityProviderResponse smart constructor.

Constructors

UpdateCapacityProviderResponse'

Fields

capacityProvider :: Maybe CapacityProvider

Details about the capacity provider.

httpStatus :: Int

The response's http status code.

newUpdateCapacityProviderResponse

Arguments

:: Int
$sel:httpStatus:UpdateCapacityProviderResponse'

-> UpdateCapacityProviderResponse

Create a value of UpdateCapacityProviderResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:capacityProvider:UpdateCapacityProviderResponse', updateCapacityProviderResponse_capacityProvider - Details about the capacity provider.

$sel:httpStatus:UpdateCapacityProviderResponse', updateCapacityProviderResponse_httpStatus - The response's http status code.

Response Lenses

updateCapacityProviderResponse_capacityProvider :: Lens' UpdateCapacityProviderResponse (Maybe CapacityProvider)

Details about the capacity provider.

updateCapacityProviderResponse_httpStatus :: Lens' UpdateCapacityProviderResponse Int

The response's http status code.