Network.AWS.ECS.UpdateCluster

Description

Updates the cluster.

Creating a Request

data UpdateCluster

See: newUpdateCluster smart constructor.

Constructors

UpdateCluster'

Fields

configuration :: Maybe ClusterConfiguration

The execute command configuration for the cluster.

settings :: Maybe [ClusterSetting]

The cluster settings for your cluster.

cluster :: Text

The name of the cluster to modify the settings for.

newUpdateCluster

Arguments

:: Text
$sel:cluster:UpdateCluster'

-> UpdateCluster

Create a value of UpdateCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configuration:UpdateCluster', updateCluster_configuration - The execute command configuration for the cluster.

$sel:settings:UpdateCluster', updateCluster_settings - The cluster settings for your cluster.

$sel:cluster:UpdateCluster', updateCluster_cluster - The name of the cluster to modify the settings for.

Request Lenses

updateCluster_configuration :: Lens' UpdateCluster (Maybe ClusterConfiguration)

The execute command configuration for the cluster.

updateCluster_settings :: Lens' UpdateCluster (Maybe [ClusterSetting])

The cluster settings for your cluster.

updateCluster_cluster :: Lens' UpdateCluster Text

The name of the cluster to modify the settings for.

Destructuring the Response

data UpdateClusterResponse

See: newUpdateClusterResponse smart constructor.

Constructors

UpdateClusterResponse'

Fields

cluster :: Maybe Cluster

Details about the cluster.

httpStatus :: Int

The response's http status code.

newUpdateClusterResponse

Arguments

:: Int
$sel:httpStatus:UpdateClusterResponse'

-> UpdateClusterResponse

Create a value of UpdateClusterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cluster:UpdateCluster', updateClusterResponse_cluster - Details about the cluster.

$sel:httpStatus:UpdateClusterResponse', updateClusterResponse_httpStatus - The response's http status code.

Response Lenses

updateClusterResponse_cluster :: Lens' UpdateClusterResponse (Maybe Cluster)

Details about the cluster.

updateClusterResponse_httpStatus :: Lens' UpdateClusterResponse Int

The response's http status code.