Network.AWS.ECS.UpdateContainerInstancesState

Description

Modifies the status of an Amazon ECS container instance.

Once a container instance has reached an ACTIVE state, you can change the status of a container instance to DRAINING to manually remove an instance from a cluster, for example to perform system updates, update the Docker daemon, or scale down the cluster size.

A container instance cannot be changed to DRAINING until it has reached an ACTIVE status. If the instance is in any other status, an error will be received.

When you set a container instance to DRAINING, Amazon ECS prevents new tasks from being scheduled for placement on the container instance and replacement service tasks are started on other container instances in the cluster if the resources are available. Service tasks on the container instance that are in the PENDING state are stopped immediately.

Service tasks on the container instance that are in the RUNNING state are stopped and replaced according to the service's deployment configuration parameters, minimumHealthyPercent and maximumPercent. You can change the deployment configuration of your service using UpdateService.

If minimumHealthyPercent is below 100%, the scheduler can ignore desiredCount temporarily during task replacement. For example, desiredCount is four tasks, a minimum of 50% allows the scheduler to stop two existing tasks before starting two new tasks. If the minimum is 100%, the service scheduler can't remove existing tasks until the replacement tasks are considered healthy. Tasks for services that do not use a load balancer are considered healthy if they are in the RUNNING state. Tasks for services that use a load balancer are considered healthy if they are in the RUNNING state and the container instance they are hosted on is reported as healthy by the load balancer.

The maximumPercent parameter represents an upper limit on the number of running tasks during task replacement, which enables you to define the replacement batch size. For example, if desiredCount is four tasks, a maximum of 200% starts four new tasks before stopping the four tasks to be drained, provided that the cluster resources required to do this are available. If the maximum is 100%, then replacement tasks can't start until the draining tasks have stopped.

Any PENDING or RUNNING tasks that do not belong to a service are not affected. You must wait for them to finish or stop them manually.

A container instance has completed draining when it has no more RUNNING tasks. You can verify this using ListTasks.

When a container instance has been drained, you can set a container instance to ACTIVE status and once it has reached that status the Amazon ECS scheduler can begin scheduling tasks on the instance again.

Creating a Request

data UpdateContainerInstancesState

See: newUpdateContainerInstancesState smart constructor.

Constructors

UpdateContainerInstancesState'

Fields

cluster :: Maybe Text

The short name or full Amazon Resource Name (ARN) of the cluster that hosts the container instance to update. If you do not specify a cluster, the default cluster is assumed.

containerInstances :: [Text]

A list of container instance IDs or full ARN entries.

status :: ContainerInstanceStatus

The container instance state with which to update the container instance. The only valid values for this action are ACTIVE and DRAINING. A container instance can only be updated to DRAINING status once it has reached an ACTIVE state. If a container instance is in REGISTERING, DEREGISTERING, or REGISTRATION_FAILED state you can describe the container instance but will be unable to update the container instance state.

newUpdateContainerInstancesState

Arguments

:: ContainerInstanceStatus
$sel:status:UpdateContainerInstancesState'

-> UpdateContainerInstancesState

Create a value of UpdateContainerInstancesState with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cluster:UpdateContainerInstancesState', updateContainerInstancesState_cluster - The short name or full Amazon Resource Name (ARN) of the cluster that hosts the container instance to update. If you do not specify a cluster, the default cluster is assumed.

$sel:containerInstances:UpdateContainerInstancesState', updateContainerInstancesState_containerInstances - A list of container instance IDs or full ARN entries.

$sel:status:UpdateContainerInstancesState', updateContainerInstancesState_status - The container instance state with which to update the container instance. The only valid values for this action are ACTIVE and DRAINING. A container instance can only be updated to DRAINING status once it has reached an ACTIVE state. If a container instance is in REGISTERING, DEREGISTERING, or REGISTRATION_FAILED state you can describe the container instance but will be unable to update the container instance state.

Request Lenses

updateContainerInstancesState_cluster :: Lens' UpdateContainerInstancesState (Maybe Text)

The short name or full Amazon Resource Name (ARN) of the cluster that hosts the container instance to update. If you do not specify a cluster, the default cluster is assumed.

updateContainerInstancesState_containerInstances :: Lens' UpdateContainerInstancesState [Text]

A list of container instance IDs or full ARN entries.

updateContainerInstancesState_status :: Lens' UpdateContainerInstancesState ContainerInstanceStatus

The container instance state with which to update the container instance. The only valid values for this action are ACTIVE and DRAINING. A container instance can only be updated to DRAINING status once it has reached an ACTIVE state. If a container instance is in REGISTERING, DEREGISTERING, or REGISTRATION_FAILED state you can describe the container instance but will be unable to update the container instance state.

Destructuring the Response

data UpdateContainerInstancesStateResponse

See: newUpdateContainerInstancesStateResponse smart constructor.

Constructors

UpdateContainerInstancesStateResponse'

Fields

failures :: Maybe [Failure]

Any failures associated with the call.

containerInstances :: Maybe [ContainerInstance]

The list of container instances.

httpStatus :: Int

The response's http status code.

newUpdateContainerInstancesStateResponse

Arguments

:: Int
$sel:httpStatus:UpdateContainerInstancesStateResponse'

-> UpdateContainerInstancesStateResponse

Create a value of UpdateContainerInstancesStateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failures:UpdateContainerInstancesStateResponse', updateContainerInstancesStateResponse_failures - Any failures associated with the call.

$sel:containerInstances:UpdateContainerInstancesState', updateContainerInstancesStateResponse_containerInstances - The list of container instances.

$sel:httpStatus:UpdateContainerInstancesStateResponse', updateContainerInstancesStateResponse_httpStatus - The response's http status code.

Response Lenses

updateContainerInstancesStateResponse_failures :: Lens' UpdateContainerInstancesStateResponse (Maybe [Failure])

Any failures associated with the call.

updateContainerInstancesStateResponse_containerInstances :: Lens' UpdateContainerInstancesStateResponse (Maybe [ContainerInstance])

The list of container instances.

updateContainerInstancesStateResponse_httpStatus :: Lens' UpdateContainerInstancesStateResponse Int

The response's http status code.