Network.AWS.ECS.UpdateServicePrimaryTaskSet

Description

Modifies which task set in a service is the primary task set. Any parameters that are updated on the primary task set in a service will transition to the service. This is used when a service uses the EXTERNAL deployment controller type. For more information, see Amazon ECS Deployment Types in the Amazon Elastic Container Service Developer Guide.

Creating a Request

data UpdateServicePrimaryTaskSet

See: newUpdateServicePrimaryTaskSet smart constructor.

Constructors

UpdateServicePrimaryTaskSet'

Fields

cluster :: Text

The short name or full Amazon Resource Name (ARN) of the cluster that hosts the service that the task set exists in.

service :: Text

The short name or full Amazon Resource Name (ARN) of the service that the task set exists in.

primaryTaskSet :: Text

The short name or full Amazon Resource Name (ARN) of the task set to set as the primary task set in the deployment.

newUpdateServicePrimaryTaskSet

Arguments

:: Text
$sel:cluster:UpdateServicePrimaryTaskSet'

-> Text
$sel:service:UpdateServicePrimaryTaskSet'

-> Text
$sel:primaryTaskSet:UpdateServicePrimaryTaskSet'

-> UpdateServicePrimaryTaskSet

Create a value of UpdateServicePrimaryTaskSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cluster:UpdateServicePrimaryTaskSet', updateServicePrimaryTaskSet_cluster - The short name or full Amazon Resource Name (ARN) of the cluster that hosts the service that the task set exists in.

$sel:service:UpdateServicePrimaryTaskSet', updateServicePrimaryTaskSet_service - The short name or full Amazon Resource Name (ARN) of the service that the task set exists in.

$sel:primaryTaskSet:UpdateServicePrimaryTaskSet', updateServicePrimaryTaskSet_primaryTaskSet - The short name or full Amazon Resource Name (ARN) of the task set to set as the primary task set in the deployment.

Request Lenses

updateServicePrimaryTaskSet_cluster :: Lens' UpdateServicePrimaryTaskSet Text

The short name or full Amazon Resource Name (ARN) of the cluster that hosts the service that the task set exists in.

updateServicePrimaryTaskSet_service :: Lens' UpdateServicePrimaryTaskSet Text

The short name or full Amazon Resource Name (ARN) of the service that the task set exists in.

updateServicePrimaryTaskSet_primaryTaskSet :: Lens' UpdateServicePrimaryTaskSet Text

The short name or full Amazon Resource Name (ARN) of the task set to set as the primary task set in the deployment.

Destructuring the Response

data UpdateServicePrimaryTaskSetResponse

See: newUpdateServicePrimaryTaskSetResponse smart constructor.

Constructors

UpdateServicePrimaryTaskSetResponse'

Fields

taskSet :: Maybe TaskSet

Details about the task set.

httpStatus :: Int

The response's http status code.

newUpdateServicePrimaryTaskSetResponse

Arguments

:: Int
$sel:httpStatus:UpdateServicePrimaryTaskSetResponse'

-> UpdateServicePrimaryTaskSetResponse

Create a value of UpdateServicePrimaryTaskSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:taskSet:UpdateServicePrimaryTaskSetResponse', updateServicePrimaryTaskSetResponse_taskSet - Details about the task set.

$sel:httpStatus:UpdateServicePrimaryTaskSetResponse', updateServicePrimaryTaskSetResponse_httpStatus - The response's http status code.

Response Lenses

updateServicePrimaryTaskSetResponse_taskSet :: Lens' UpdateServicePrimaryTaskSetResponse (Maybe TaskSet)

Details about the task set.

updateServicePrimaryTaskSetResponse_httpStatus :: Lens' UpdateServicePrimaryTaskSetResponse Int

The response's http status code.