Network.AWS.EFS.CreateAccessPoint

Description

Creates an EFS access point. An access point is an application-specific view into an EFS file system that applies an operating system user and group, and a file system path, to any file system request made through the access point. The operating system user and group override any identity information provided by the NFS client. The file system path is exposed as the access point's root directory. Applications using the access point can only access data in its own directory and below. To learn more, see Mounting a file system using EFS access points.

This operation requires permissions for the elasticfilesystem:CreateAccessPoint action.

Creating a Request

data CreateAccessPoint

See: newCreateAccessPoint smart constructor.

Constructors

CreateAccessPoint'

Fields

rootDirectory :: Maybe RootDirectory

Specifies the directory on the Amazon EFS file system that the access point exposes as the root directory of your file system to NFS clients using the access point. The clients using the access point can only access the root directory and below. If the RootDirectory > Path specified does not exist, EFS creates it and applies the CreationInfo settings when a client connects to an access point. When specifying a RootDirectory, you need to provide the Path, and the CreationInfo.

Amazon EFS creates a root directory only if you have provided the CreationInfo: OwnUid, OwnGID, and permissions for the directory. If you do not provide this information, Amazon EFS does not create the root directory. If the root directory does not exist, attempts to mount using the access point will fail.

posixUser :: Maybe PosixUser

The operating system user and group applied to all file system requests made using the access point.

tags :: Maybe [Tag]

Creates tags associated with the access point. Each tag is a key-value pair, each key must be unique. For more information, see Tagging Amazon Web Services resources in the Amazon Web Services General Reference Guide.

clientToken :: Text

A string of up to 64 ASCII characters that Amazon EFS uses to ensure idempotent creation.

fileSystemId :: Text

The ID of the EFS file system that the access point provides access to.

newCreateAccessPoint

Arguments

:: Text
$sel:clientToken:CreateAccessPoint'

-> Text
$sel:fileSystemId:CreateAccessPoint'

-> CreateAccessPoint

Create a value of CreateAccessPoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:rootDirectory:CreateAccessPoint', createAccessPoint_rootDirectory - Specifies the directory on the Amazon EFS file system that the access point exposes as the root directory of your file system to NFS clients using the access point. The clients using the access point can only access the root directory and below. If the RootDirectory > Path specified does not exist, EFS creates it and applies the CreationInfo settings when a client connects to an access point. When specifying a RootDirectory, you need to provide the Path, and the CreationInfo.

Amazon EFS creates a root directory only if you have provided the CreationInfo: OwnUid, OwnGID, and permissions for the directory. If you do not provide this information, Amazon EFS does not create the root directory. If the root directory does not exist, attempts to mount using the access point will fail.

$sel:posixUser:CreateAccessPoint', createAccessPoint_posixUser - The operating system user and group applied to all file system requests made using the access point.

$sel:tags:CreateAccessPoint', createAccessPoint_tags - Creates tags associated with the access point. Each tag is a key-value pair, each key must be unique. For more information, see Tagging Amazon Web Services resources in the Amazon Web Services General Reference Guide.

$sel:clientToken:CreateAccessPoint', createAccessPoint_clientToken - A string of up to 64 ASCII characters that Amazon EFS uses to ensure idempotent creation.

$sel:fileSystemId:CreateAccessPoint', createAccessPoint_fileSystemId - The ID of the EFS file system that the access point provides access to.

Request Lenses

createAccessPoint_rootDirectory :: Lens' CreateAccessPoint (Maybe RootDirectory)

Specifies the directory on the Amazon EFS file system that the access point exposes as the root directory of your file system to NFS clients using the access point. The clients using the access point can only access the root directory and below. If the RootDirectory > Path specified does not exist, EFS creates it and applies the CreationInfo settings when a client connects to an access point. When specifying a RootDirectory, you need to provide the Path, and the CreationInfo.

Amazon EFS creates a root directory only if you have provided the CreationInfo: OwnUid, OwnGID, and permissions for the directory. If you do not provide this information, Amazon EFS does not create the root directory. If the root directory does not exist, attempts to mount using the access point will fail.

createAccessPoint_posixUser :: Lens' CreateAccessPoint (Maybe PosixUser)

The operating system user and group applied to all file system requests made using the access point.

createAccessPoint_tags :: Lens' CreateAccessPoint (Maybe [Tag])

Creates tags associated with the access point. Each tag is a key-value pair, each key must be unique. For more information, see Tagging Amazon Web Services resources in the Amazon Web Services General Reference Guide.

createAccessPoint_clientToken :: Lens' CreateAccessPoint Text

A string of up to 64 ASCII characters that Amazon EFS uses to ensure idempotent creation.

createAccessPoint_fileSystemId :: Lens' CreateAccessPoint Text

The ID of the EFS file system that the access point provides access to.

Destructuring the Response

data AccessPointDescription

Provides a description of an EFS file system access point.

See: newAccessPointDescription smart constructor.

Constructors

AccessPointDescription'

Fields

ownerId :: Maybe Text

Identified the Amazon Web Services account that owns the access point resource.

accessPointId :: Maybe Text

The ID of the access point, assigned by Amazon EFS.

accessPointArn :: Maybe Text

The unique Amazon Resource Name (ARN) associated with the access point.

name :: Maybe Text

The name of the access point. This is the value of the Name tag.

rootDirectory :: Maybe RootDirectory

The directory on the Amazon EFS file system that the access point exposes as the root directory to NFS clients using the access point.

posixUser :: Maybe PosixUser

The full POSIX identity, including the user ID, group ID, and secondary group IDs on the access point that is used for all file operations by NFS clients using the access point.

tags :: Maybe [Tag]

The tags associated with the access point, presented as an array of Tag objects.

lifeCycleState :: Maybe LifeCycleState

Identifies the lifecycle phase of the access point.

fileSystemId :: Maybe Text

The ID of the EFS file system that the access point applies to.

clientToken :: Maybe Text

The opaque string specified in the request to ensure idempotent creation.

newAccessPointDescription :: AccessPointDescription

Create a value of AccessPointDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ownerId:AccessPointDescription', accessPointDescription_ownerId - Identified the Amazon Web Services account that owns the access point resource.

$sel:accessPointId:AccessPointDescription', accessPointDescription_accessPointId - The ID of the access point, assigned by Amazon EFS.

$sel:accessPointArn:AccessPointDescription', accessPointDescription_accessPointArn - The unique Amazon Resource Name (ARN) associated with the access point.

$sel:name:AccessPointDescription', accessPointDescription_name - The name of the access point. This is the value of the Name tag.

$sel:rootDirectory:AccessPointDescription', accessPointDescription_rootDirectory - The directory on the Amazon EFS file system that the access point exposes as the root directory to NFS clients using the access point.

$sel:posixUser:AccessPointDescription', accessPointDescription_posixUser - The full POSIX identity, including the user ID, group ID, and secondary group IDs on the access point that is used for all file operations by NFS clients using the access point.

$sel:tags:AccessPointDescription', accessPointDescription_tags - The tags associated with the access point, presented as an array of Tag objects.

$sel:lifeCycleState:AccessPointDescription', accessPointDescription_lifeCycleState - Identifies the lifecycle phase of the access point.

$sel:fileSystemId:AccessPointDescription', accessPointDescription_fileSystemId - The ID of the EFS file system that the access point applies to.

$sel:clientToken:AccessPointDescription', accessPointDescription_clientToken - The opaque string specified in the request to ensure idempotent creation.

Response Lenses

accessPointDescription_ownerId :: Lens' AccessPointDescription (Maybe Text)

Identified the Amazon Web Services account that owns the access point resource.

accessPointDescription_accessPointId :: Lens' AccessPointDescription (Maybe Text)

The ID of the access point, assigned by Amazon EFS.

accessPointDescription_accessPointArn :: Lens' AccessPointDescription (Maybe Text)

The unique Amazon Resource Name (ARN) associated with the access point.

accessPointDescription_name :: Lens' AccessPointDescription (Maybe Text)

The name of the access point. This is the value of the Name tag.

accessPointDescription_rootDirectory :: Lens' AccessPointDescription (Maybe RootDirectory)

The directory on the Amazon EFS file system that the access point exposes as the root directory to NFS clients using the access point.

accessPointDescription_posixUser :: Lens' AccessPointDescription (Maybe PosixUser)

The full POSIX identity, including the user ID, group ID, and secondary group IDs on the access point that is used for all file operations by NFS clients using the access point.

accessPointDescription_tags :: Lens' AccessPointDescription (Maybe [Tag])

The tags associated with the access point, presented as an array of Tag objects.

accessPointDescription_lifeCycleState :: Lens' AccessPointDescription (Maybe LifeCycleState)

Identifies the lifecycle phase of the access point.

accessPointDescription_fileSystemId :: Lens' AccessPointDescription (Maybe Text)

The ID of the EFS file system that the access point applies to.

accessPointDescription_clientToken :: Lens' AccessPointDescription (Maybe Text)

The opaque string specified in the request to ensure idempotent creation.