Network.AWS.EFS.CreateMountTarget

Description

Creates a mount target for a file system. You can then mount the file system on EC2 instances by using the mount target.

You can create one mount target in each Availability Zone in your VPC. All EC2 instances in a VPC within a given Availability Zone share a single mount target for a given file system. If you have multiple subnets in an Availability Zone, you create a mount target in one of the subnets. EC2 instances do not need to be in the same subnet as the mount target in order to access their file system.

You can create only one mount target for an EFS file system using One Zone storage classes. You must create that mount target in the same Availability Zone in which the file system is located. Use the AvailabilityZoneName and AvailabiltyZoneId properties in the DescribeFileSystems response object to get this information. Use the subnetId associated with the file system's Availability Zone when creating the mount target.

For more information, see Amazon EFS: How it Works.

To create a mount target for a file system, the file system's lifecycle state must be available. For more information, see DescribeFileSystems.

In the request, provide the following:

The file system ID for which you are creating the mount target.

A subnet ID, which determines the following:

The VPC in which Amazon EFS creates the mount target

The Availability Zone in which Amazon EFS creates the mount target

The IP address range from which Amazon EFS selects the IP address of the mount target (if you don't specify an IP address in the request)

After creating the mount target, Amazon EFS returns a response that includes, a MountTargetId and an IpAddress. You use this IP address when mounting the file system in an EC2 instance. You can also use the mount target's DNS name when mounting the file system. The EC2 instance on which you mount the file system by using the mount target can resolve the mount target's DNS name to its IP address. For more information, see How it Works: Implementation Overview.

Note that you can create mount targets for a file system in only one VPC, and there can be only one mount target per Availability Zone. That is, if the file system already has one or more mount targets created for it, the subnet specified in the request to add another mount target must meet the following requirements:

Must belong to the same VPC as the subnets of the existing mount targets

Must not be in the same Availability Zone as any of the subnets of the existing mount targets

If the request satisfies the requirements, Amazon EFS does the following:

Creates a new mount target in the specified subnet.

Also creates a new network interface in the subnet as follows:

If the request provides an IpAddress, Amazon EFS assigns that IP address to the network interface. Otherwise, Amazon EFS assigns a free address in the subnet (in the same way that the Amazon EC2 CreateNetworkInterface call does when a request does not specify a primary private IP address).

If the request provides SecurityGroups, this network interface is associated with those security groups. Otherwise, it belongs to the default security group for the subnet's VPC.

Assigns the description Mount target fsmt-id for file system fs-id where fsmt-id is the mount target ID, and fs-id is the FileSystemId.

Sets the requesterManaged property of the network interface to true, and the requesterId value to EFS.

Each Amazon EFS mount target has one corresponding requester-managed EC2 network interface. After the network interface is created, Amazon EFS sets the NetworkInterfaceId field in the mount target's description to the network interface ID, and the IpAddress field to its address. If network interface creation fails, the entire CreateMountTarget operation fails.

The CreateMountTarget call returns only after creating the network interface, but while the mount target state is still creating, you can check the mount target creation status by calling the DescribeMountTargets operation, which among other things returns the mount target state.

We recommend that you create a mount target in each of the Availability Zones. There are cost considerations for using a file system in an Availability Zone through a mount target created in another Availability Zone. For more information, see Amazon EFS. In addition, by always using a mount target local to the instance's Availability Zone, you eliminate a partial failure scenario. If the Availability Zone in which your mount target is created goes down, then you can't access your file system through that mount target.

This operation requires permissions for the following action on the file system:

elasticfilesystem:CreateMountTarget

This operation also requires permissions for the following Amazon EC2 actions:

ec2:DescribeSubnets

ec2:DescribeNetworkInterfaces

ec2:CreateNetworkInterface

Creating a Request

data CreateMountTarget

See: newCreateMountTarget smart constructor.

Constructors

CreateMountTarget'

Fields

securityGroups :: Maybe [Text]

Up to five VPC security group IDs, of the form sg-xxxxxxxx. These must be for the same VPC as subnet specified.

ipAddress :: Maybe Text

Valid IPv4 address within the address range of the specified subnet.

fileSystemId :: Text

The ID of the file system for which to create the mount target.

subnetId :: Text

The ID of the subnet to add the mount target in. For file systems that use One Zone storage classes, use the subnet that is associated with the file system's Availability Zone.

newCreateMountTarget

Arguments

:: Text
$sel:fileSystemId:CreateMountTarget'

-> Text
$sel:subnetId:CreateMountTarget'

-> CreateMountTarget

Create a value of CreateMountTarget with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:securityGroups:CreateMountTarget', createMountTarget_securityGroups - Up to five VPC security group IDs, of the form sg-xxxxxxxx. These must be for the same VPC as subnet specified.

$sel:ipAddress:CreateMountTarget', createMountTarget_ipAddress - Valid IPv4 address within the address range of the specified subnet.

$sel:fileSystemId:CreateMountTarget', createMountTarget_fileSystemId - The ID of the file system for which to create the mount target.

$sel:subnetId:CreateMountTarget', createMountTarget_subnetId - The ID of the subnet to add the mount target in. For file systems that use One Zone storage classes, use the subnet that is associated with the file system's Availability Zone.

Request Lenses

createMountTarget_securityGroups :: Lens' CreateMountTarget (Maybe [Text])

Up to five VPC security group IDs, of the form sg-xxxxxxxx. These must be for the same VPC as subnet specified.

createMountTarget_ipAddress :: Lens' CreateMountTarget (Maybe Text)

Valid IPv4 address within the address range of the specified subnet.

createMountTarget_fileSystemId :: Lens' CreateMountTarget Text

The ID of the file system for which to create the mount target.

createMountTarget_subnetId :: Lens' CreateMountTarget Text

The ID of the subnet to add the mount target in. For file systems that use One Zone storage classes, use the subnet that is associated with the file system's Availability Zone.

Destructuring the Response

data MountTargetDescription

Provides a description of a mount target.

See: newMountTargetDescription smart constructor.

Constructors

MountTargetDescription'

Fields

ownerId :: Maybe Text

Amazon Web Services account ID that owns the resource.

availabilityZoneName :: Maybe Text

The name of the Availability Zone in which the mount target is located. Availability Zones are independently mapped to names for each Amazon Web Services account. For example, the Availability Zone us-east-1a for your Amazon Web Services account might not be the same location as us-east-1a for another Amazon Web Services account.

availabilityZoneId :: Maybe Text

The unique and consistent identifier of the Availability Zone that the mount target resides in. For example, use1-az1 is an AZ ID for the us-east-1 Region and it has the same location in every Amazon Web Services account.

ipAddress :: Maybe Text

Address at which the file system can be mounted by using the mount target.

networkInterfaceId :: Maybe Text

The ID of the network interface that Amazon EFS created when it created the mount target.

vpcId :: Maybe Text

The virtual private cloud (VPC) ID that the mount target is configured in.

mountTargetId :: Text

System-assigned mount target ID.

fileSystemId :: Text

The ID of the file system for which the mount target is intended.

subnetId :: Text

The ID of the mount target's subnet.

lifeCycleState :: LifeCycleState

Lifecycle state of the mount target.

newMountTargetDescription

Arguments

:: Text
$sel:mountTargetId:MountTargetDescription'

-> Text
$sel:fileSystemId:MountTargetDescription'

-> Text
$sel:subnetId:MountTargetDescription'

-> LifeCycleState
$sel:lifeCycleState:MountTargetDescription'

-> MountTargetDescription

Create a value of MountTargetDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ownerId:MountTargetDescription', mountTargetDescription_ownerId - Amazon Web Services account ID that owns the resource.

$sel:availabilityZoneName:MountTargetDescription', mountTargetDescription_availabilityZoneName - The name of the Availability Zone in which the mount target is located. Availability Zones are independently mapped to names for each Amazon Web Services account. For example, the Availability Zone us-east-1a for your Amazon Web Services account might not be the same location as us-east-1a for another Amazon Web Services account.

$sel:availabilityZoneId:MountTargetDescription', mountTargetDescription_availabilityZoneId - The unique and consistent identifier of the Availability Zone that the mount target resides in. For example, use1-az1 is an AZ ID for the us-east-1 Region and it has the same location in every Amazon Web Services account.

$sel:ipAddress:MountTargetDescription', mountTargetDescription_ipAddress - Address at which the file system can be mounted by using the mount target.

$sel:networkInterfaceId:MountTargetDescription', mountTargetDescription_networkInterfaceId - The ID of the network interface that Amazon EFS created when it created the mount target.

$sel:vpcId:MountTargetDescription', mountTargetDescription_vpcId - The virtual private cloud (VPC) ID that the mount target is configured in.

$sel:mountTargetId:MountTargetDescription', mountTargetDescription_mountTargetId - System-assigned mount target ID.

$sel:fileSystemId:MountTargetDescription', mountTargetDescription_fileSystemId - The ID of the file system for which the mount target is intended.

$sel:subnetId:MountTargetDescription', mountTargetDescription_subnetId - The ID of the mount target's subnet.

$sel:lifeCycleState:MountTargetDescription', mountTargetDescription_lifeCycleState - Lifecycle state of the mount target.

Response Lenses

mountTargetDescription_ownerId :: Lens' MountTargetDescription (Maybe Text)

Amazon Web Services account ID that owns the resource.

mountTargetDescription_availabilityZoneName :: Lens' MountTargetDescription (Maybe Text)

The name of the Availability Zone in which the mount target is located. Availability Zones are independently mapped to names for each Amazon Web Services account. For example, the Availability Zone us-east-1a for your Amazon Web Services account might not be the same location as us-east-1a for another Amazon Web Services account.

mountTargetDescription_availabilityZoneId :: Lens' MountTargetDescription (Maybe Text)

The unique and consistent identifier of the Availability Zone that the mount target resides in. For example, use1-az1 is an AZ ID for the us-east-1 Region and it has the same location in every Amazon Web Services account.

mountTargetDescription_ipAddress :: Lens' MountTargetDescription (Maybe Text)

Address at which the file system can be mounted by using the mount target.

mountTargetDescription_networkInterfaceId :: Lens' MountTargetDescription (Maybe Text)

The ID of the network interface that Amazon EFS created when it created the mount target.

mountTargetDescription_vpcId :: Lens' MountTargetDescription (Maybe Text)

The virtual private cloud (VPC) ID that the mount target is configured in.

mountTargetDescription_mountTargetId :: Lens' MountTargetDescription Text

System-assigned mount target ID.

mountTargetDescription_fileSystemId :: Lens' MountTargetDescription Text

The ID of the file system for which the mount target is intended.

mountTargetDescription_subnetId :: Lens' MountTargetDescription Text

The ID of the mount target's subnet.

mountTargetDescription_lifeCycleState :: Lens' MountTargetDescription LifeCycleState

Lifecycle state of the mount target.