Network.AWS.EFS.DescribeAccountPreferences

Description

Returns the account preferences settings for the Amazon Web Services account associated with the user making the request, in the current Amazon Web Services Region. For more information, see Managing Amazon EFS resource IDs.

Creating a Request

data DescribeAccountPreferences

See: newDescribeAccountPreferences smart constructor.

Constructors

DescribeAccountPreferences'

Fields

nextToken :: Maybe Text

(Optional) You can use NextToken in a subsequent request to fetch the next page of Amazon Web Services account preferences if the response payload was paginated.

maxResults :: Maybe Natural

(Optional) When retrieving account preferences, you can optionally specify the MaxItems parameter to limit the number of objects returned in a response. The default value is 100.

newDescribeAccountPreferences :: DescribeAccountPreferences

Create a value of DescribeAccountPreferences with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeAccountPreferences', describeAccountPreferences_nextToken - (Optional) You can use NextToken in a subsequent request to fetch the next page of Amazon Web Services account preferences if the response payload was paginated.

$sel:maxResults:DescribeAccountPreferences', describeAccountPreferences_maxResults - (Optional) When retrieving account preferences, you can optionally specify the MaxItems parameter to limit the number of objects returned in a response. The default value is 100.

Request Lenses

describeAccountPreferences_nextToken :: Lens' DescribeAccountPreferences (Maybe Text)

(Optional) You can use NextToken in a subsequent request to fetch the next page of Amazon Web Services account preferences if the response payload was paginated.

describeAccountPreferences_maxResults :: Lens' DescribeAccountPreferences (Maybe Natural)

(Optional) When retrieving account preferences, you can optionally specify the MaxItems parameter to limit the number of objects returned in a response. The default value is 100.

Destructuring the Response

data DescribeAccountPreferencesResponse

See: newDescribeAccountPreferencesResponse smart constructor.

Constructors

DescribeAccountPreferencesResponse'

Fields

resourceIdPreference :: Maybe ResourceIdPreference

Describes the resource ID preference setting for the Amazon Web Services account associated with the user making the request, in the current Amazon Web Services Region.

nextToken :: Maybe Text

Present if there are more records than returned in the response. You can use the NextToken in the subsequent request to fetch the additional descriptions.

httpStatus :: Int

The response's http status code.

newDescribeAccountPreferencesResponse

Arguments

:: Int
$sel:httpStatus:DescribeAccountPreferencesResponse'

-> DescribeAccountPreferencesResponse

Create a value of DescribeAccountPreferencesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceIdPreference:DescribeAccountPreferencesResponse', describeAccountPreferencesResponse_resourceIdPreference - Describes the resource ID preference setting for the Amazon Web Services account associated with the user making the request, in the current Amazon Web Services Region.

$sel:nextToken:DescribeAccountPreferences', describeAccountPreferencesResponse_nextToken - Present if there are more records than returned in the response. You can use the NextToken in the subsequent request to fetch the additional descriptions.

$sel:httpStatus:DescribeAccountPreferencesResponse', describeAccountPreferencesResponse_httpStatus - The response's http status code.

Response Lenses

describeAccountPreferencesResponse_resourceIdPreference :: Lens' DescribeAccountPreferencesResponse (Maybe ResourceIdPreference)

Describes the resource ID preference setting for the Amazon Web Services account associated with the user making the request, in the current Amazon Web Services Region.

describeAccountPreferencesResponse_nextToken :: Lens' DescribeAccountPreferencesResponse (Maybe Text)

Present if there are more records than returned in the response. You can use the NextToken in the subsequent request to fetch the additional descriptions.

describeAccountPreferencesResponse_httpStatus :: Lens' DescribeAccountPreferencesResponse Int

The response's http status code.