Network.AWS.EFS.DescribeFileSystems

Description

Returns the description of a specific Amazon EFS file system if either the file system CreationToken or the FileSystemId is provided. Otherwise, it returns descriptions of all file systems owned by the caller's Amazon Web Services account in the Amazon Web Services Region of the endpoint that you're calling.

When retrieving all file system descriptions, you can optionally specify the MaxItems parameter to limit the number of descriptions in a response. Currently, this number is automatically set to 10. If more file system descriptions remain, Amazon EFS returns a NextMarker, an opaque token, in the response. In this case, you should send a subsequent request with the Marker request parameter set to the value of NextMarker.

To retrieve a list of your file system descriptions, this operation is used in an iterative process, where DescribeFileSystems is called first without the Marker and then the operation continues to call it with the Marker parameter set to the value of the NextMarker from the previous response until the response has no NextMarker.

The order of file systems returned in the response of one DescribeFileSystems call and the order of file systems returned across the responses of a multi-call iteration is unspecified.

This operation requires permissions for the elasticfilesystem:DescribeFileSystems action.

This operation returns paginated results.

Creating a Request

data DescribeFileSystems

See: newDescribeFileSystems smart constructor.

Constructors

DescribeFileSystems'

Fields

creationToken :: Maybe Text

(Optional) Restricts the list to the file system with this creation token (String). You specify a creation token when you create an Amazon EFS file system.

fileSystemId :: Maybe Text

(Optional) ID of the file system whose description you want to retrieve (String).

maxItems :: Maybe Natural

(Optional) Specifies the maximum number of file systems to return in the response (integer). This number is automatically set to 100. The response is paginated at 100 per page if you have more than 100 file systems.

marker :: Maybe Text

(Optional) Opaque pagination token returned from a previous DescribeFileSystems operation (String). If present, specifies to continue the list from where the returning call had left off.

newDescribeFileSystems :: DescribeFileSystems

Create a value of DescribeFileSystems with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationToken:DescribeFileSystems', describeFileSystems_creationToken - (Optional) Restricts the list to the file system with this creation token (String). You specify a creation token when you create an Amazon EFS file system.

$sel:fileSystemId:DescribeFileSystems', describeFileSystems_fileSystemId - (Optional) ID of the file system whose description you want to retrieve (String).

$sel:maxItems:DescribeFileSystems', describeFileSystems_maxItems - (Optional) Specifies the maximum number of file systems to return in the response (integer). This number is automatically set to 100. The response is paginated at 100 per page if you have more than 100 file systems.

$sel:marker:DescribeFileSystems', describeFileSystems_marker - (Optional) Opaque pagination token returned from a previous DescribeFileSystems operation (String). If present, specifies to continue the list from where the returning call had left off.

Request Lenses

describeFileSystems_creationToken :: Lens' DescribeFileSystems (Maybe Text)

(Optional) Restricts the list to the file system with this creation token (String). You specify a creation token when you create an Amazon EFS file system.

describeFileSystems_fileSystemId :: Lens' DescribeFileSystems (Maybe Text)

(Optional) ID of the file system whose description you want to retrieve (String).

describeFileSystems_maxItems :: Lens' DescribeFileSystems (Maybe Natural)

(Optional) Specifies the maximum number of file systems to return in the response (integer). This number is automatically set to 100. The response is paginated at 100 per page if you have more than 100 file systems.

describeFileSystems_marker :: Lens' DescribeFileSystems (Maybe Text)

(Optional) Opaque pagination token returned from a previous DescribeFileSystems operation (String). If present, specifies to continue the list from where the returning call had left off.

Destructuring the Response

data DescribeFileSystemsResponse

See: newDescribeFileSystemsResponse smart constructor.

Constructors

DescribeFileSystemsResponse'

Fields

nextMarker :: Maybe Text

Present if there are more file systems than returned in the response (String). You can use the NextMarker in the subsequent request to fetch the descriptions.

fileSystems :: Maybe [FileSystemDescription]

An array of file system descriptions.

marker :: Maybe Text

Present if provided by caller in the request (String).

httpStatus :: Int

The response's http status code.

newDescribeFileSystemsResponse

Arguments

:: Int
$sel:httpStatus:DescribeFileSystemsResponse'

-> DescribeFileSystemsResponse

Create a value of DescribeFileSystemsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextMarker:DescribeFileSystemsResponse', describeFileSystemsResponse_nextMarker - Present if there are more file systems than returned in the response (String). You can use the NextMarker in the subsequent request to fetch the descriptions.

$sel:fileSystems:DescribeFileSystemsResponse', describeFileSystemsResponse_fileSystems - An array of file system descriptions.

$sel:marker:DescribeFileSystems', describeFileSystemsResponse_marker - Present if provided by caller in the request (String).

$sel:httpStatus:DescribeFileSystemsResponse', describeFileSystemsResponse_httpStatus - The response's http status code.

Response Lenses

describeFileSystemsResponse_nextMarker :: Lens' DescribeFileSystemsResponse (Maybe Text)

Present if there are more file systems than returned in the response (String). You can use the NextMarker in the subsequent request to fetch the descriptions.

describeFileSystemsResponse_fileSystems :: Lens' DescribeFileSystemsResponse (Maybe [FileSystemDescription])

An array of file system descriptions.

describeFileSystemsResponse_marker :: Lens' DescribeFileSystemsResponse (Maybe Text)

Present if provided by caller in the request (String).

describeFileSystemsResponse_httpStatus :: Lens' DescribeFileSystemsResponse Int

The response's http status code.