Network.AWS.EFS.PutLifecycleConfiguration

Description

Enables lifecycle management by creating a new LifecycleConfiguration object. A LifecycleConfiguration object defines when files in an Amazon EFS file system are automatically transitioned to the lower-cost EFS Infrequent Access (IA) storage class. To enable EFS Intelligent Tiering, set the value of TransitionToPrimaryStorageClass to AFTER_1_ACCESS. For more information, see EFS Lifecycle Management.

A LifecycleConfiguration applies to all files in a file system.

Each Amazon EFS file system supports one lifecycle configuration, which applies to all files in the file system. If a LifecycleConfiguration object already exists for the specified file system, a PutLifecycleConfiguration call modifies the existing configuration. A PutLifecycleConfiguration call with an empty LifecyclePolicies array in the request body deletes any existing LifecycleConfiguration and disables lifecycle management.

In the request, specify the following:

The ID for the file system for which you are enabling, disabling, or modifying lifecycle management.

A LifecyclePolicies array of LifecyclePolicy objects that define when files are moved to the IA storage class. The array can contain only one LifecyclePolicy item.

This operation requires permissions for the elasticfilesystem:PutLifecycleConfiguration operation.

To apply a LifecycleConfiguration object to an encrypted file system, you need the same Key Management Service permissions as when you created the encrypted file system.

Creating a Request

data PutLifecycleConfiguration

See: newPutLifecycleConfiguration smart constructor.

Constructors

PutLifecycleConfiguration'

Fields

fileSystemId :: Text

The ID of the file system for which you are creating the LifecycleConfiguration object (String).

lifecyclePolicies :: [LifecyclePolicy]

An array of LifecyclePolicy objects that define the file system's LifecycleConfiguration object. A LifecycleConfiguration object tells lifecycle management when to transition files from the Standard storage class to the Infrequent Access storage class.

newPutLifecycleConfiguration

Arguments

:: Text
$sel:fileSystemId:PutLifecycleConfiguration'

-> PutLifecycleConfiguration

Create a value of PutLifecycleConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fileSystemId:PutLifecycleConfiguration', putLifecycleConfiguration_fileSystemId - The ID of the file system for which you are creating the LifecycleConfiguration object (String).

$sel:lifecyclePolicies:PutLifecycleConfiguration', putLifecycleConfiguration_lifecyclePolicies - An array of LifecyclePolicy objects that define the file system's LifecycleConfiguration object. A LifecycleConfiguration object tells lifecycle management when to transition files from the Standard storage class to the Infrequent Access storage class.

Request Lenses

putLifecycleConfiguration_fileSystemId :: Lens' PutLifecycleConfiguration Text

The ID of the file system for which you are creating the LifecycleConfiguration object (String).

putLifecycleConfiguration_lifecyclePolicies :: Lens' PutLifecycleConfiguration [LifecyclePolicy]

An array of LifecyclePolicy objects that define the file system's LifecycleConfiguration object. A LifecycleConfiguration object tells lifecycle management when to transition files from the Standard storage class to the Infrequent Access storage class.

Destructuring the Response

data LifecycleConfigurationDescription

See: newLifecycleConfigurationDescription smart constructor.

Constructors

LifecycleConfigurationDescription'

Fields

lifecyclePolicies :: Maybe [LifecyclePolicy]

An array of lifecycle management policies. EFS supports a maximum of one policy per file system.

newLifecycleConfigurationDescription :: LifecycleConfigurationDescription

Create a value of LifecycleConfigurationDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lifecyclePolicies:LifecycleConfigurationDescription', lifecycleConfigurationDescription_lifecyclePolicies - An array of lifecycle management policies. EFS supports a maximum of one policy per file system.

Response Lenses

lifecycleConfigurationDescription_lifecyclePolicies :: Lens' LifecycleConfigurationDescription (Maybe [LifecyclePolicy])

An array of lifecycle management policies. EFS supports a maximum of one policy per file system.