Network.AWS.EFS.Types

Description

Service Configuration

defaultService :: Service

API version 2015-02-01 of the Amazon Elastic File System SDK configuration.

Errors

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

Returned if the Backup service is not available in the Amazon Web Services Region in which the request was made.

_MountTargetNotFound :: AsError a => Getting (First ServiceError) a ServiceError

Returned if there is no mount target with the specified ID found in the caller's Amazon Web Services account.

_SecurityGroupLimitExceeded :: AsError a => Getting (First ServiceError) a ServiceError

Returned if the size of SecurityGroups specified in the request is greater than five.

_SecurityGroupNotFound :: AsError a => Getting (First ServiceError) a ServiceError

Returned if one of the specified security groups doesn't exist in the subnet's VPC.

_MountTargetConflict :: AsError a => Getting (First ServiceError) a ServiceError

Returned if the mount target would violate one of the specified restrictions based on the file system's existing mount targets.

_UnsupportedAvailabilityZone :: AsError a => Getting (First ServiceError) a ServiceError

Returned if the requested Amazon EFS functionality is not available in the specified Availability Zone.

_FileSystemLimitExceeded :: AsError a => Getting (First ServiceError) a ServiceError

Returned if the Amazon Web Services account has already created the maximum number of file systems allowed per account.

_TooManyRequests :: AsError a => Getting (First ServiceError) a ServiceError

Returned if you don’t wait at least 24 hours before changing the throughput mode, or decreasing the Provisioned Throughput value.

_NetworkInterfaceLimitExceeded :: AsError a => Getting (First ServiceError) a ServiceError

The calling account has reached the limit for elastic network interfaces for the specific Amazon Web Services Region. The client should try to delete some elastic network interfaces or get the account limit raised. For more information, see Amazon VPC Limits in the Amazon VPC User Guide (see the Network interfaces per VPC entry in the table).

_FileSystemAlreadyExists :: AsError a => Getting (First ServiceError) a ServiceError

Returned if the file system you are trying to create already exists, with the creation token you provided.

_SubnetNotFound :: AsError a => Getting (First ServiceError) a ServiceError

Returned if there is no subnet with ID SubnetId provided in the request.

_AvailabilityZonesMismatch :: AsError a => Getting (First ServiceError) a ServiceError

Returned if the Availability Zone that was specified for a mount target is different from the Availability Zone that was specified for One Zone storage classes. For more information, see Regional and One Zone storage redundancy.

_FileSystemNotFound :: AsError a => Getting (First ServiceError) a ServiceError

Returned if the specified FileSystemId value doesn't exist in the requester's Amazon Web Services account.

_IncorrectFileSystemLifeCycleState :: AsError a => Getting (First ServiceError) a ServiceError

Returned if the file system's lifecycle state is not "available".

_BadRequest :: AsError a => Getting (First ServiceError) a ServiceError

Returned if the request is malformed or contains an error such as an invalid parameter value or a missing required parameter.

_NoFreeAddressesInSubnet :: AsError a => Getting (First ServiceError) a ServiceError

Returned if IpAddress was not specified in the request and there are no free IP addresses in the subnet.

_ThroughputLimitExceeded :: AsError a => Getting (First ServiceError) a ServiceError

Returned if the throughput mode or amount of provisioned throughput can't be changed because the throughput limit of 1024 MiB/s has been reached.

_DependencyTimeout :: AsError a => Getting (First ServiceError) a ServiceError

The service timed out trying to fulfill the request, and the client should try the call again.

_FileSystemInUse :: AsError a => Getting (First ServiceError) a ServiceError

Returned if a file system has mount targets.

_IncorrectMountTargetState :: AsError a => Getting (First ServiceError) a ServiceError

Returned if the mount target is not in the correct state for the operation.

_InternalServerError :: AsError a => Getting (First ServiceError) a ServiceError

Returned if an error occurred on the server side.

_IpAddressInUse :: AsError a => Getting (First ServiceError) a ServiceError

Returned if the request specified an IpAddress that is already in use in the subnet.

_PolicyNotFound :: AsError a => Getting (First ServiceError) a ServiceError

Returned if the default file system policy is in effect for the EFS file system specified.

_AccessPointNotFound :: AsError a => Getting (First ServiceError) a ServiceError

Returned if the specified AccessPointId value doesn't exist in the requester's Amazon Web Services account.

_InsufficientThroughputCapacity :: AsError a => Getting (First ServiceError) a ServiceError

Returned if there's not enough capacity to provision additional throughput. This value might be returned when you try to create a file system in provisioned throughput mode, when you attempt to increase the provisioned throughput of an existing file system, or when you attempt to change an existing file system from bursting to provisioned throughput mode. Try again later.

_InvalidPolicyException :: AsError a => Getting (First ServiceError) a ServiceError

Returned if the FileSystemPolicy is is malformed or contains an error such as an invalid parameter value or a missing required parameter. Returned in the case of a policy lockout safety check error.

_AccessPointAlreadyExists :: AsError a => Getting (First ServiceError) a ServiceError

Returned if the access point you are trying to create already exists, with the creation token you provided in the request.

_AccessPointLimitExceeded :: AsError a => Getting (First ServiceError) a ServiceError

Returned if the Amazon Web Services account has already created the maximum number of access points allowed per file system.

BackupStatus

newtype BackupStatus

Constructors

BackupStatus'

Fields

fromBackupStatus :: Text

Bundled Patterns

pattern BackupStatus_DISABLED :: BackupStatus

pattern BackupStatus_DISABLING :: BackupStatus

pattern BackupStatus_ENABLED :: BackupStatus

pattern BackupStatus_ENABLING :: BackupStatus

LifeCycleState

newtype LifeCycleState

Constructors

LifeCycleState'

Fields

fromLifeCycleState :: Text

Bundled Patterns

pattern LifeCycleState_Available :: LifeCycleState

pattern LifeCycleState_Creating :: LifeCycleState

pattern LifeCycleState_Deleted :: LifeCycleState

pattern LifeCycleState_Deleting :: LifeCycleState

pattern LifeCycleState_Error :: LifeCycleState

pattern LifeCycleState_Updating :: LifeCycleState

PerformanceMode

newtype PerformanceMode

Constructors

PerformanceMode'

Fields

fromPerformanceMode :: Text

Bundled Patterns

pattern PerformanceMode_GeneralPurpose :: PerformanceMode

pattern PerformanceMode_MaxIO :: PerformanceMode

Resource

newtype Resource

An EFS resource, for example a file system or a mount target.

Constructors

Resource'

Fields

fromResource :: Text

Bundled Patterns

pattern Resource_FILE_SYSTEM :: Resource

pattern Resource_MOUNT_TARGET :: Resource

ResourceIdType

newtype ResourceIdType

A preference indicating a choice to use 63bit/32bit IDs for all applicable resources.

Constructors

ResourceIdType'

Fields

fromResourceIdType :: Text

Bundled Patterns

pattern ResourceIdType_LONG_ID :: ResourceIdType

pattern ResourceIdType_SHORT_ID :: ResourceIdType

ThroughputMode

newtype ThroughputMode

Constructors

ThroughputMode'

Fields

fromThroughputMode :: Text

Bundled Patterns

pattern ThroughputMode_Bursting :: ThroughputMode

pattern ThroughputMode_Provisioned :: ThroughputMode

TransitionToIARules

newtype TransitionToIARules

Constructors

TransitionToIARules'

Fields

fromTransitionToIARules :: Text

Bundled Patterns

pattern TransitionToIARules_AFTER_14_DAYS :: TransitionToIARules

pattern TransitionToIARules_AFTER_30_DAYS :: TransitionToIARules

pattern TransitionToIARules_AFTER_60_DAYS :: TransitionToIARules

pattern TransitionToIARules_AFTER_7_DAYS :: TransitionToIARules

pattern TransitionToIARules_AFTER_90_DAYS :: TransitionToIARules

TransitionToPrimaryStorageClassRules

newtype TransitionToPrimaryStorageClassRules

Constructors

TransitionToPrimaryStorageClassRules'

Fields

fromTransitionToPrimaryStorageClassRules :: Text

Bundled Patterns

pattern TransitionToPrimaryStorageClassRules_AFTER_1_ACCESS :: TransitionToPrimaryStorageClassRules

AccessPointDescription

data AccessPointDescription

Provides a description of an EFS file system access point.

See: newAccessPointDescription smart constructor.

Constructors

AccessPointDescription'

Fields

posixUser :: Maybe PosixUser

The full POSIX identity, including the user ID, group ID, and secondary group IDs on the access point that is used for all file operations by NFS clients using the access point.

rootDirectory :: Maybe RootDirectory

The directory on the Amazon EFS file system that the access point exposes as the root directory to NFS clients using the access point.

clientToken :: Maybe Text

The opaque string specified in the request to ensure idempotent creation.

accessPointId :: Maybe Text

The ID of the access point, assigned by Amazon EFS.

fileSystemId :: Maybe Text

The ID of the EFS file system that the access point applies to.

ownerId :: Maybe Text

Identified the Amazon Web Services account that owns the access point resource.

name :: Maybe Text

The name of the access point. This is the value of the Name tag.

accessPointArn :: Maybe Text

The unique Amazon Resource Name (ARN) associated with the access point.

lifeCycleState :: Maybe LifeCycleState

Identifies the lifecycle phase of the access point.

tags :: Maybe [Tag]

The tags associated with the access point, presented as an array of Tag objects.

newAccessPointDescription :: AccessPointDescription

Create a value of AccessPointDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:posixUser:AccessPointDescription', accessPointDescription_posixUser - The full POSIX identity, including the user ID, group ID, and secondary group IDs on the access point that is used for all file operations by NFS clients using the access point.

$sel:rootDirectory:AccessPointDescription', accessPointDescription_rootDirectory - The directory on the Amazon EFS file system that the access point exposes as the root directory to NFS clients using the access point.

$sel:clientToken:AccessPointDescription', accessPointDescription_clientToken - The opaque string specified in the request to ensure idempotent creation.

$sel:accessPointId:AccessPointDescription', accessPointDescription_accessPointId - The ID of the access point, assigned by Amazon EFS.

$sel:fileSystemId:AccessPointDescription', accessPointDescription_fileSystemId - The ID of the EFS file system that the access point applies to.

$sel:ownerId:AccessPointDescription', accessPointDescription_ownerId - Identified the Amazon Web Services account that owns the access point resource.

$sel:name:AccessPointDescription', accessPointDescription_name - The name of the access point. This is the value of the Name tag.

$sel:accessPointArn:AccessPointDescription', accessPointDescription_accessPointArn - The unique Amazon Resource Name (ARN) associated with the access point.

$sel:lifeCycleState:AccessPointDescription', accessPointDescription_lifeCycleState - Identifies the lifecycle phase of the access point.

$sel:tags:AccessPointDescription', accessPointDescription_tags - The tags associated with the access point, presented as an array of Tag objects.

accessPointDescription_posixUser :: Lens' AccessPointDescription (Maybe PosixUser)

The full POSIX identity, including the user ID, group ID, and secondary group IDs on the access point that is used for all file operations by NFS clients using the access point.

accessPointDescription_rootDirectory :: Lens' AccessPointDescription (Maybe RootDirectory)

The directory on the Amazon EFS file system that the access point exposes as the root directory to NFS clients using the access point.

accessPointDescription_clientToken :: Lens' AccessPointDescription (Maybe Text)

The opaque string specified in the request to ensure idempotent creation.

accessPointDescription_accessPointId :: Lens' AccessPointDescription (Maybe Text)

The ID of the access point, assigned by Amazon EFS.

accessPointDescription_fileSystemId :: Lens' AccessPointDescription (Maybe Text)

The ID of the EFS file system that the access point applies to.

accessPointDescription_ownerId :: Lens' AccessPointDescription (Maybe Text)

Identified the Amazon Web Services account that owns the access point resource.

accessPointDescription_name :: Lens' AccessPointDescription (Maybe Text)

The name of the access point. This is the value of the Name tag.

accessPointDescription_accessPointArn :: Lens' AccessPointDescription (Maybe Text)

The unique Amazon Resource Name (ARN) associated with the access point.

accessPointDescription_lifeCycleState :: Lens' AccessPointDescription (Maybe LifeCycleState)

Identifies the lifecycle phase of the access point.

accessPointDescription_tags :: Lens' AccessPointDescription (Maybe [Tag])

The tags associated with the access point, presented as an array of Tag objects.

BackupPolicy

data BackupPolicy

The backup policy for the file system used to create automatic daily backups. If status has a value of ENABLED, the file system is being automatically backed up. For more information, see Automatic backups.

See: newBackupPolicy smart constructor.

Constructors

BackupPolicy'

Fields

status :: BackupStatus

Describes the status of the file system's backup policy.

ENABLED - EFS is automatically backing up the file system.

ENABLING - EFS is turning on automatic backups for the file system.

DISABLED - automatic back ups are turned off for the file system.

DISABLING - EFS is turning off automatic backups for the file system.

newBackupPolicy

Arguments

:: BackupStatus
$sel:status:BackupPolicy'

-> BackupPolicy

Create a value of BackupPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:BackupPolicy', backupPolicy_status - Describes the status of the file system's backup policy.

ENABLED - EFS is automatically backing up the file system.

ENABLING - EFS is turning on automatic backups for the file system.

DISABLED - automatic back ups are turned off for the file system.

DISABLING - EFS is turning off automatic backups for the file system.

backupPolicy_status :: Lens' BackupPolicy BackupStatus

Describes the status of the file system's backup policy.

ENABLED - EFS is automatically backing up the file system.

ENABLING - EFS is turning on automatic backups for the file system.

DISABLED - automatic back ups are turned off for the file system.

DISABLING - EFS is turning off automatic backups for the file system.

BackupPolicyDescription

data BackupPolicyDescription

See: newBackupPolicyDescription smart constructor.

Constructors

BackupPolicyDescription'

Fields

backupPolicy :: Maybe BackupPolicy

Describes the file system's backup policy, indicating whether automatic backups are turned on or off..

newBackupPolicyDescription :: BackupPolicyDescription

Create a value of BackupPolicyDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:backupPolicy:BackupPolicyDescription', backupPolicyDescription_backupPolicy - Describes the file system's backup policy, indicating whether automatic backups are turned on or off..

backupPolicyDescription_backupPolicy :: Lens' BackupPolicyDescription (Maybe BackupPolicy)

Describes the file system's backup policy, indicating whether automatic backups are turned on or off..

CreationInfo

data CreationInfo

Required if the RootDirectory > Path specified does not exist. Specifies the POSIX IDs and permissions to apply to the access point's RootDirectory > Path. If the access point root directory does not exist, EFS creates it with these settings when a client connects to the access point. When specifying CreationInfo, you must include values for all properties.

Amazon EFS creates a root directory only if you have provided the CreationInfo: OwnUid, OwnGID, and permissions for the directory. If you do not provide this information, Amazon EFS does not create the root directory. If the root directory does not exist, attempts to mount using the access point will fail.

If you do not provide CreationInfo and the specified RootDirectory does not exist, attempts to mount the file system using the access point will fail.

See: newCreationInfo smart constructor.

Constructors

CreationInfo'

Fields

ownerUid :: Natural

Specifies the POSIX user ID to apply to the RootDirectory. Accepts values from 0 to 2^32 (4294967295).

ownerGid :: Natural

Specifies the POSIX group ID to apply to the RootDirectory. Accepts values from 0 to 2^32 (4294967295).

permissions :: Text

Specifies the POSIX permissions to apply to the RootDirectory, in the format of an octal number representing the file's mode bits.

newCreationInfo

Arguments

:: Natural
$sel:ownerUid:CreationInfo'

-> Natural
$sel:ownerGid:CreationInfo'

-> Text
$sel:permissions:CreationInfo'

-> CreationInfo

Create a value of CreationInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ownerUid:CreationInfo', creationInfo_ownerUid - Specifies the POSIX user ID to apply to the RootDirectory. Accepts values from 0 to 2^32 (4294967295).

$sel:ownerGid:CreationInfo', creationInfo_ownerGid - Specifies the POSIX group ID to apply to the RootDirectory. Accepts values from 0 to 2^32 (4294967295).

$sel:permissions:CreationInfo', creationInfo_permissions - Specifies the POSIX permissions to apply to the RootDirectory, in the format of an octal number representing the file's mode bits.

creationInfo_ownerUid :: Lens' CreationInfo Natural

Specifies the POSIX user ID to apply to the RootDirectory. Accepts values from 0 to 2^32 (4294967295).

creationInfo_ownerGid :: Lens' CreationInfo Natural

Specifies the POSIX group ID to apply to the RootDirectory. Accepts values from 0 to 2^32 (4294967295).

creationInfo_permissions :: Lens' CreationInfo Text

Specifies the POSIX permissions to apply to the RootDirectory, in the format of an octal number representing the file's mode bits.

FileSystemDescription

data FileSystemDescription

A description of the file system.

See: newFileSystemDescription smart constructor.

Constructors

FileSystemDescription'

Fields

availabilityZoneId :: Maybe Text

The unique and consistent identifier of the Availability Zone in which the file system's One Zone storage classes exist. For example, use1-az1 is an Availability Zone ID for the us-east-1 Amazon Web Services Region, and it has the same location in every Amazon Web Services account.

provisionedThroughputInMibps :: Maybe Double

The amount of provisioned throughput, measured in MiB/s, for the file system. Valid for file systems using ThroughputMode set to provisioned.

availabilityZoneName :: Maybe Text

Describes the Amazon Web Services Availability Zone in which the file system is located, and is valid only for file systems using One Zone storage classes. For more information, see Using EFS storage classes in the Amazon EFS User Guide.

fileSystemArn :: Maybe Text

The Amazon Resource Name (ARN) for the EFS file system, in the format arn:aws:elasticfilesystem:region:account-id:file-system/file-system-id . Example with sample data: arn:aws:elasticfilesystem:us-west-2:1111333322228888:file-system/fs-01234567

encrypted :: Maybe Bool

A Boolean value that, if true, indicates that the file system is encrypted.

throughputMode :: Maybe ThroughputMode

Displays the file system's throughput mode. For more information, see Throughput modes in the Amazon EFS User Guide.

kmsKeyId :: Maybe Text

The ID of an Key Management Service customer master key (CMK) that was used to protect the encrypted file system.

name :: Maybe Text

You can add tags to a file system, including a Name tag. For more information, see CreateFileSystem. If the file system has a Name tag, Amazon EFS returns the value in this field.

ownerId :: Text

The Amazon Web Services account that created the file system. If the file system was created by an IAM user, the parent account to which the user belongs is the owner.

creationToken :: Text

The opaque string specified in the request.

fileSystemId :: Text

The ID of the file system, assigned by Amazon EFS.

creationTime :: POSIX

The time that the file system was created, in seconds (since 1970-01-01T00:00:00Z).

lifeCycleState :: LifeCycleState

The lifecycle phase of the file system.

numberOfMountTargets :: Natural

The current number of mount targets that the file system has. For more information, see CreateMountTarget.

sizeInBytes :: FileSystemSize

The latest known metered size (in bytes) of data stored in the file system, in its Value field, and the time at which that size was determined in its Timestamp field. The Timestamp value is the integer number of seconds since 1970-01-01T00:00:00Z. The SizeInBytes value doesn't represent the size of a consistent snapshot of the file system, but it is eventually consistent when there are no writes to the file system. That is, SizeInBytes represents actual size only if the file system is not modified for a period longer than a couple of hours. Otherwise, the value is not the exact size that the file system was at any point in time.

performanceMode :: PerformanceMode

The performance mode of the file system.

tags :: [Tag]

The tags associated with the file system, presented as an array of Tag objects.

newFileSystemDescription

Arguments

:: Text
$sel:ownerId:FileSystemDescription'

-> Text
$sel:creationToken:FileSystemDescription'

-> Text
$sel:fileSystemId:FileSystemDescription'

-> UTCTime
$sel:creationTime:FileSystemDescription'

-> LifeCycleState
$sel:lifeCycleState:FileSystemDescription'

-> Natural
$sel:numberOfMountTargets:FileSystemDescription'

-> FileSystemSize
$sel:sizeInBytes:FileSystemDescription'

-> PerformanceMode
$sel:performanceMode:FileSystemDescription'

-> FileSystemDescription

Create a value of FileSystemDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:availabilityZoneId:FileSystemDescription', fileSystemDescription_availabilityZoneId - The unique and consistent identifier of the Availability Zone in which the file system's One Zone storage classes exist. For example, use1-az1 is an Availability Zone ID for the us-east-1 Amazon Web Services Region, and it has the same location in every Amazon Web Services account.

$sel:provisionedThroughputInMibps:FileSystemDescription', fileSystemDescription_provisionedThroughputInMibps - The amount of provisioned throughput, measured in MiB/s, for the file system. Valid for file systems using ThroughputMode set to provisioned.

$sel:availabilityZoneName:FileSystemDescription', fileSystemDescription_availabilityZoneName - Describes the Amazon Web Services Availability Zone in which the file system is located, and is valid only for file systems using One Zone storage classes. For more information, see Using EFS storage classes in the Amazon EFS User Guide.

$sel:fileSystemArn:FileSystemDescription', fileSystemDescription_fileSystemArn - The Amazon Resource Name (ARN) for the EFS file system, in the format arn:aws:elasticfilesystem:region:account-id:file-system/file-system-id . Example with sample data: arn:aws:elasticfilesystem:us-west-2:1111333322228888:file-system/fs-01234567

$sel:encrypted:FileSystemDescription', fileSystemDescription_encrypted - A Boolean value that, if true, indicates that the file system is encrypted.

$sel:throughputMode:FileSystemDescription', fileSystemDescription_throughputMode - Displays the file system's throughput mode. For more information, see Throughput modes in the Amazon EFS User Guide.

$sel:kmsKeyId:FileSystemDescription', fileSystemDescription_kmsKeyId - The ID of an Key Management Service customer master key (CMK) that was used to protect the encrypted file system.

$sel:name:FileSystemDescription', fileSystemDescription_name - You can add tags to a file system, including a Name tag. For more information, see CreateFileSystem. If the file system has a Name tag, Amazon EFS returns the value in this field.

$sel:ownerId:FileSystemDescription', fileSystemDescription_ownerId - The Amazon Web Services account that created the file system. If the file system was created by an IAM user, the parent account to which the user belongs is the owner.

$sel:creationToken:FileSystemDescription', fileSystemDescription_creationToken - The opaque string specified in the request.

$sel:fileSystemId:FileSystemDescription', fileSystemDescription_fileSystemId - The ID of the file system, assigned by Amazon EFS.

$sel:creationTime:FileSystemDescription', fileSystemDescription_creationTime - The time that the file system was created, in seconds (since 1970-01-01T00:00:00Z).

$sel:lifeCycleState:FileSystemDescription', fileSystemDescription_lifeCycleState - The lifecycle phase of the file system.

$sel:numberOfMountTargets:FileSystemDescription', fileSystemDescription_numberOfMountTargets - The current number of mount targets that the file system has. For more information, see CreateMountTarget.

$sel:sizeInBytes:FileSystemDescription', fileSystemDescription_sizeInBytes - The latest known metered size (in bytes) of data stored in the file system, in its Value field, and the time at which that size was determined in its Timestamp field. The Timestamp value is the integer number of seconds since 1970-01-01T00:00:00Z. The SizeInBytes value doesn't represent the size of a consistent snapshot of the file system, but it is eventually consistent when there are no writes to the file system. That is, SizeInBytes represents actual size only if the file system is not modified for a period longer than a couple of hours. Otherwise, the value is not the exact size that the file system was at any point in time.

$sel:performanceMode:FileSystemDescription', fileSystemDescription_performanceMode - The performance mode of the file system.

$sel:tags:FileSystemDescription', fileSystemDescription_tags - The tags associated with the file system, presented as an array of Tag objects.

fileSystemDescription_availabilityZoneId :: Lens' FileSystemDescription (Maybe Text)

The unique and consistent identifier of the Availability Zone in which the file system's One Zone storage classes exist. For example, use1-az1 is an Availability Zone ID for the us-east-1 Amazon Web Services Region, and it has the same location in every Amazon Web Services account.

fileSystemDescription_provisionedThroughputInMibps :: Lens' FileSystemDescription (Maybe Double)

The amount of provisioned throughput, measured in MiB/s, for the file system. Valid for file systems using ThroughputMode set to provisioned.

fileSystemDescription_availabilityZoneName :: Lens' FileSystemDescription (Maybe Text)

Describes the Amazon Web Services Availability Zone in which the file system is located, and is valid only for file systems using One Zone storage classes. For more information, see Using EFS storage classes in the Amazon EFS User Guide.

fileSystemDescription_fileSystemArn :: Lens' FileSystemDescription (Maybe Text)

The Amazon Resource Name (ARN) for the EFS file system, in the format arn:aws:elasticfilesystem:region:account-id:file-system/file-system-id . Example with sample data: arn:aws:elasticfilesystem:us-west-2:1111333322228888:file-system/fs-01234567

fileSystemDescription_encrypted :: Lens' FileSystemDescription (Maybe Bool)

A Boolean value that, if true, indicates that the file system is encrypted.

fileSystemDescription_throughputMode :: Lens' FileSystemDescription (Maybe ThroughputMode)

Displays the file system's throughput mode. For more information, see Throughput modes in the Amazon EFS User Guide.

fileSystemDescription_kmsKeyId :: Lens' FileSystemDescription (Maybe Text)

The ID of an Key Management Service customer master key (CMK) that was used to protect the encrypted file system.

fileSystemDescription_name :: Lens' FileSystemDescription (Maybe Text)

You can add tags to a file system, including a Name tag. For more information, see CreateFileSystem. If the file system has a Name tag, Amazon EFS returns the value in this field.

fileSystemDescription_ownerId :: Lens' FileSystemDescription Text

The Amazon Web Services account that created the file system. If the file system was created by an IAM user, the parent account to which the user belongs is the owner.

fileSystemDescription_creationToken :: Lens' FileSystemDescription Text

The opaque string specified in the request.

fileSystemDescription_fileSystemId :: Lens' FileSystemDescription Text

The ID of the file system, assigned by Amazon EFS.

fileSystemDescription_creationTime :: Lens' FileSystemDescription UTCTime

The time that the file system was created, in seconds (since 1970-01-01T00:00:00Z).

fileSystemDescription_lifeCycleState :: Lens' FileSystemDescription LifeCycleState

The lifecycle phase of the file system.

fileSystemDescription_numberOfMountTargets :: Lens' FileSystemDescription Natural

The current number of mount targets that the file system has. For more information, see CreateMountTarget.

fileSystemDescription_sizeInBytes :: Lens' FileSystemDescription FileSystemSize

The latest known metered size (in bytes) of data stored in the file system, in its Value field, and the time at which that size was determined in its Timestamp field. The Timestamp value is the integer number of seconds since 1970-01-01T00:00:00Z. The SizeInBytes value doesn't represent the size of a consistent snapshot of the file system, but it is eventually consistent when there are no writes to the file system. That is, SizeInBytes represents actual size only if the file system is not modified for a period longer than a couple of hours. Otherwise, the value is not the exact size that the file system was at any point in time.

fileSystemDescription_performanceMode :: Lens' FileSystemDescription PerformanceMode

The performance mode of the file system.

fileSystemDescription_tags :: Lens' FileSystemDescription [Tag]

The tags associated with the file system, presented as an array of Tag objects.

FileSystemPolicyDescription

data FileSystemPolicyDescription

See: newFileSystemPolicyDescription smart constructor.

Constructors

FileSystemPolicyDescription'

Fields

fileSystemId :: Maybe Text

Specifies the EFS file system to which the FileSystemPolicy applies.

policy :: Maybe Text

The JSON formatted FileSystemPolicy for the EFS file system.

newFileSystemPolicyDescription :: FileSystemPolicyDescription

Create a value of FileSystemPolicyDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fileSystemId:FileSystemPolicyDescription', fileSystemPolicyDescription_fileSystemId - Specifies the EFS file system to which the FileSystemPolicy applies.

$sel:policy:FileSystemPolicyDescription', fileSystemPolicyDescription_policy - The JSON formatted FileSystemPolicy for the EFS file system.

fileSystemPolicyDescription_fileSystemId :: Lens' FileSystemPolicyDescription (Maybe Text)

Specifies the EFS file system to which the FileSystemPolicy applies.

fileSystemPolicyDescription_policy :: Lens' FileSystemPolicyDescription (Maybe Text)

The JSON formatted FileSystemPolicy for the EFS file system.

FileSystemSize

data FileSystemSize

The latest known metered size (in bytes) of data stored in the file system, in its Value field, and the time at which that size was determined in its Timestamp field. The value doesn't represent the size of a consistent snapshot of the file system, but it is eventually consistent when there are no writes to the file system. That is, the value represents the actual size only if the file system is not modified for a period longer than a couple of hours. Otherwise, the value is not necessarily the exact size the file system was at any instant in time.

See: newFileSystemSize smart constructor.

Constructors

FileSystemSize'

Fields

valueInIA :: Maybe Natural

The latest known metered size (in bytes) of data stored in the Infrequent Access storage class.

valueInStandard :: Maybe Natural

The latest known metered size (in bytes) of data stored in the Standard storage class.

timestamp :: Maybe POSIX

The time at which the size of data, returned in the Value field, was determined. The value is the integer number of seconds since 1970-01-01T00:00:00Z.

value :: Natural

The latest known metered size (in bytes) of data stored in the file system.

newFileSystemSize

Arguments

:: Natural
$sel:value:FileSystemSize'

-> FileSystemSize

Create a value of FileSystemSize with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:valueInIA:FileSystemSize', fileSystemSize_valueInIA - The latest known metered size (in bytes) of data stored in the Infrequent Access storage class.

$sel:valueInStandard:FileSystemSize', fileSystemSize_valueInStandard - The latest known metered size (in bytes) of data stored in the Standard storage class.

$sel:timestamp:FileSystemSize', fileSystemSize_timestamp - The time at which the size of data, returned in the Value field, was determined. The value is the integer number of seconds since 1970-01-01T00:00:00Z.

$sel:value:FileSystemSize', fileSystemSize_value - The latest known metered size (in bytes) of data stored in the file system.

fileSystemSize_valueInIA :: Lens' FileSystemSize (Maybe Natural)

The latest known metered size (in bytes) of data stored in the Infrequent Access storage class.

fileSystemSize_valueInStandard :: Lens' FileSystemSize (Maybe Natural)

The latest known metered size (in bytes) of data stored in the Standard storage class.

fileSystemSize_timestamp :: Lens' FileSystemSize (Maybe UTCTime)

The time at which the size of data, returned in the Value field, was determined. The value is the integer number of seconds since 1970-01-01T00:00:00Z.

fileSystemSize_value :: Lens' FileSystemSize Natural

The latest known metered size (in bytes) of data stored in the file system.

LifecycleConfigurationDescription

data LifecycleConfigurationDescription

See: newLifecycleConfigurationDescription smart constructor.

Constructors

LifecycleConfigurationDescription'

Fields

lifecyclePolicies :: Maybe [LifecyclePolicy]

An array of lifecycle management policies. EFS supports a maximum of one policy per file system.

newLifecycleConfigurationDescription :: LifecycleConfigurationDescription

Create a value of LifecycleConfigurationDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lifecyclePolicies:LifecycleConfigurationDescription', lifecycleConfigurationDescription_lifecyclePolicies - An array of lifecycle management policies. EFS supports a maximum of one policy per file system.

lifecycleConfigurationDescription_lifecyclePolicies :: Lens' LifecycleConfigurationDescription (Maybe [LifecyclePolicy])

An array of lifecycle management policies. EFS supports a maximum of one policy per file system.

LifecyclePolicy

data LifecyclePolicy

Describes a policy used by EFS lifecycle management and EFS intelligent tiering that specifies when to transition files into and out of the file system's Infrequent Access (IA) storage class. For more information, see EFS Intelligent‐Tiering and EFS Lifecycle Management.

When using the put-lifecycle-configuration CLI command or the PutLifecycleConfiguration API action, Amazon EFS requires that each LifecyclePolicy object have only a single transition. This means that in a request body, LifecyclePolicies needs to be structured as an array of LifecyclePolicy objects, one object for each transition, TransitionToIA, TransitionToPrimaryStorageClass. For more information, see the request examples in PutLifecycleConfiguration.

See: newLifecyclePolicy smart constructor.

Constructors

LifecyclePolicy'

Fields

transitionToIA :: Maybe TransitionToIARules

Describes the period of time that a file is not accessed, after which it transitions to IA storage. Metadata operations such as listing the contents of a directory don't count as file access events.

transitionToPrimaryStorageClass :: Maybe TransitionToPrimaryStorageClassRules

Describes when to transition a file from IA storage to primary storage. Metadata operations such as listing the contents of a directory don't count as file access events.

newLifecyclePolicy :: LifecyclePolicy

Create a value of LifecyclePolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:transitionToIA:LifecyclePolicy', lifecyclePolicy_transitionToIA - Describes the period of time that a file is not accessed, after which it transitions to IA storage. Metadata operations such as listing the contents of a directory don't count as file access events.

$sel:transitionToPrimaryStorageClass:LifecyclePolicy', lifecyclePolicy_transitionToPrimaryStorageClass - Describes when to transition a file from IA storage to primary storage. Metadata operations such as listing the contents of a directory don't count as file access events.

lifecyclePolicy_transitionToIA :: Lens' LifecyclePolicy (Maybe TransitionToIARules)

Describes the period of time that a file is not accessed, after which it transitions to IA storage. Metadata operations such as listing the contents of a directory don't count as file access events.

lifecyclePolicy_transitionToPrimaryStorageClass :: Lens' LifecyclePolicy (Maybe TransitionToPrimaryStorageClassRules)

Describes when to transition a file from IA storage to primary storage. Metadata operations such as listing the contents of a directory don't count as file access events.

MountTargetDescription

data MountTargetDescription

Provides a description of a mount target.

See: newMountTargetDescription smart constructor.

Constructors

MountTargetDescription'

Fields

ipAddress :: Maybe Text

Address at which the file system can be mounted by using the mount target.

availabilityZoneId :: Maybe Text

The unique and consistent identifier of the Availability Zone that the mount target resides in. For example, use1-az1 is an AZ ID for the us-east-1 Region and it has the same location in every Amazon Web Services account.

vpcId :: Maybe Text

The virtual private cloud (VPC) ID that the mount target is configured in.

availabilityZoneName :: Maybe Text

The name of the Availability Zone in which the mount target is located. Availability Zones are independently mapped to names for each Amazon Web Services account. For example, the Availability Zone us-east-1a for your Amazon Web Services account might not be the same location as us-east-1a for another Amazon Web Services account.

networkInterfaceId :: Maybe Text

The ID of the network interface that Amazon EFS created when it created the mount target.

ownerId :: Maybe Text

Amazon Web Services account ID that owns the resource.

mountTargetId :: Text

System-assigned mount target ID.

fileSystemId :: Text

The ID of the file system for which the mount target is intended.

subnetId :: Text

The ID of the mount target's subnet.

lifeCycleState :: LifeCycleState

Lifecycle state of the mount target.

newMountTargetDescription

Arguments

:: Text
$sel:mountTargetId:MountTargetDescription'

-> Text
$sel:fileSystemId:MountTargetDescription'

-> Text
$sel:subnetId:MountTargetDescription'

-> LifeCycleState
$sel:lifeCycleState:MountTargetDescription'

-> MountTargetDescription

Create a value of MountTargetDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ipAddress:MountTargetDescription', mountTargetDescription_ipAddress - Address at which the file system can be mounted by using the mount target.

$sel:availabilityZoneId:MountTargetDescription', mountTargetDescription_availabilityZoneId - The unique and consistent identifier of the Availability Zone that the mount target resides in. For example, use1-az1 is an AZ ID for the us-east-1 Region and it has the same location in every Amazon Web Services account.

$sel:vpcId:MountTargetDescription', mountTargetDescription_vpcId - The virtual private cloud (VPC) ID that the mount target is configured in.

$sel:availabilityZoneName:MountTargetDescription', mountTargetDescription_availabilityZoneName - The name of the Availability Zone in which the mount target is located. Availability Zones are independently mapped to names for each Amazon Web Services account. For example, the Availability Zone us-east-1a for your Amazon Web Services account might not be the same location as us-east-1a for another Amazon Web Services account.

$sel:networkInterfaceId:MountTargetDescription', mountTargetDescription_networkInterfaceId - The ID of the network interface that Amazon EFS created when it created the mount target.

$sel:ownerId:MountTargetDescription', mountTargetDescription_ownerId - Amazon Web Services account ID that owns the resource.

$sel:mountTargetId:MountTargetDescription', mountTargetDescription_mountTargetId - System-assigned mount target ID.

$sel:fileSystemId:MountTargetDescription', mountTargetDescription_fileSystemId - The ID of the file system for which the mount target is intended.

$sel:subnetId:MountTargetDescription', mountTargetDescription_subnetId - The ID of the mount target's subnet.

$sel:lifeCycleState:MountTargetDescription', mountTargetDescription_lifeCycleState - Lifecycle state of the mount target.

mountTargetDescription_ipAddress :: Lens' MountTargetDescription (Maybe Text)

Address at which the file system can be mounted by using the mount target.

mountTargetDescription_availabilityZoneId :: Lens' MountTargetDescription (Maybe Text)

The unique and consistent identifier of the Availability Zone that the mount target resides in. For example, use1-az1 is an AZ ID for the us-east-1 Region and it has the same location in every Amazon Web Services account.

mountTargetDescription_vpcId :: Lens' MountTargetDescription (Maybe Text)

The virtual private cloud (VPC) ID that the mount target is configured in.

mountTargetDescription_availabilityZoneName :: Lens' MountTargetDescription (Maybe Text)

The name of the Availability Zone in which the mount target is located. Availability Zones are independently mapped to names for each Amazon Web Services account. For example, the Availability Zone us-east-1a for your Amazon Web Services account might not be the same location as us-east-1a for another Amazon Web Services account.

mountTargetDescription_networkInterfaceId :: Lens' MountTargetDescription (Maybe Text)

The ID of the network interface that Amazon EFS created when it created the mount target.

mountTargetDescription_ownerId :: Lens' MountTargetDescription (Maybe Text)

Amazon Web Services account ID that owns the resource.

mountTargetDescription_mountTargetId :: Lens' MountTargetDescription Text

System-assigned mount target ID.

mountTargetDescription_fileSystemId :: Lens' MountTargetDescription Text

The ID of the file system for which the mount target is intended.

mountTargetDescription_subnetId :: Lens' MountTargetDescription Text

The ID of the mount target's subnet.

mountTargetDescription_lifeCycleState :: Lens' MountTargetDescription LifeCycleState

Lifecycle state of the mount target.

PosixUser

data PosixUser

The full POSIX identity, including the user ID, group ID, and any secondary group IDs, on the access point that is used for all file system operations performed by NFS clients using the access point.

See: newPosixUser smart constructor.

Constructors

PosixUser'

Fields

secondaryGids :: Maybe [Natural]

Secondary POSIX group IDs used for all file system operations using this access point.

uid :: Natural

The POSIX user ID used for all file system operations using this access point.

gid :: Natural

The POSIX group ID used for all file system operations using this access point.

newPosixUser

Arguments

:: Natural
$sel:uid:PosixUser'

-> Natural
$sel:gid:PosixUser'

-> PosixUser

Create a value of PosixUser with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:secondaryGids:PosixUser', posixUser_secondaryGids - Secondary POSIX group IDs used for all file system operations using this access point.

$sel:uid:PosixUser', posixUser_uid - The POSIX user ID used for all file system operations using this access point.

$sel:gid:PosixUser', posixUser_gid - The POSIX group ID used for all file system operations using this access point.

posixUser_secondaryGids :: Lens' PosixUser (Maybe [Natural])

Secondary POSIX group IDs used for all file system operations using this access point.

posixUser_uid :: Lens' PosixUser Natural

The POSIX user ID used for all file system operations using this access point.

posixUser_gid :: Lens' PosixUser Natural

The POSIX group ID used for all file system operations using this access point.

ResourceIdPreference

data ResourceIdPreference

Describes the resource type and its ID preference for the user's Amazon Web Services account, in the current Amazon Web Services Region.

See: newResourceIdPreference smart constructor.

Constructors

ResourceIdPreference'

Fields

resources :: Maybe [Resource]

Identifies the Amazon EFS resources to which the ID preference setting applies, FILE_SYSTEM and MOUNT_TARGET.

resourceIdType :: Maybe ResourceIdType

Identifies the EFS resource ID preference, either LONG_ID (17 characters) or SHORT_ID (8 characters).

newResourceIdPreference :: ResourceIdPreference

Create a value of ResourceIdPreference with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resources:ResourceIdPreference', resourceIdPreference_resources - Identifies the Amazon EFS resources to which the ID preference setting applies, FILE_SYSTEM and MOUNT_TARGET.

$sel:resourceIdType:ResourceIdPreference', resourceIdPreference_resourceIdType - Identifies the EFS resource ID preference, either LONG_ID (17 characters) or SHORT_ID (8 characters).

resourceIdPreference_resources :: Lens' ResourceIdPreference (Maybe [Resource])

Identifies the Amazon EFS resources to which the ID preference setting applies, FILE_SYSTEM and MOUNT_TARGET.

resourceIdPreference_resourceIdType :: Lens' ResourceIdPreference (Maybe ResourceIdType)

Identifies the EFS resource ID preference, either LONG_ID (17 characters) or SHORT_ID (8 characters).

RootDirectory

data RootDirectory

Specifies the directory on the Amazon EFS file system that the access point provides access to. The access point exposes the specified file system path as the root directory of your file system to applications using the access point. NFS clients using the access point can only access data in the access point's RootDirectory and it's subdirectories.

See: newRootDirectory smart constructor.

Constructors

RootDirectory'

Fields

creationInfo :: Maybe CreationInfo

(Optional) Specifies the POSIX IDs and permissions to apply to the access point's RootDirectory. If the RootDirectory > Path specified does not exist, EFS creates the root directory using the CreationInfo settings when a client connects to an access point. When specifying the CreationInfo, you must provide values for all properties.

If you do not provide CreationInfo and the specified RootDirectory > Path does not exist, attempts to mount the file system using the access point will fail.

path :: Maybe Text

Specifies the path on the EFS file system to expose as the root directory to NFS clients using the access point to access the EFS file system. A path can have up to four subdirectories. If the specified path does not exist, you are required to provide the CreationInfo.

newRootDirectory :: RootDirectory

Create a value of RootDirectory with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationInfo:RootDirectory', rootDirectory_creationInfo - (Optional) Specifies the POSIX IDs and permissions to apply to the access point's RootDirectory. If the RootDirectory > Path specified does not exist, EFS creates the root directory using the CreationInfo settings when a client connects to an access point. When specifying the CreationInfo, you must provide values for all properties.

If you do not provide CreationInfo and the specified RootDirectory > Path does not exist, attempts to mount the file system using the access point will fail.

$sel:path:RootDirectory', rootDirectory_path - Specifies the path on the EFS file system to expose as the root directory to NFS clients using the access point to access the EFS file system. A path can have up to four subdirectories. If the specified path does not exist, you are required to provide the CreationInfo.

rootDirectory_creationInfo :: Lens' RootDirectory (Maybe CreationInfo)

(Optional) Specifies the POSIX IDs and permissions to apply to the access point's RootDirectory. If the RootDirectory > Path specified does not exist, EFS creates the root directory using the CreationInfo settings when a client connects to an access point. When specifying the CreationInfo, you must provide values for all properties.

If you do not provide CreationInfo and the specified RootDirectory > Path does not exist, attempts to mount the file system using the access point will fail.

rootDirectory_path :: Lens' RootDirectory (Maybe Text)

Specifies the path on the EFS file system to expose as the root directory to NFS clients using the access point to access the EFS file system. A path can have up to four subdirectories. If the specified path does not exist, you are required to provide the CreationInfo.

Tag

data Tag

A tag is a key-value pair. Allowed characters are letters, white space, and numbers that can be represented in UTF-8, and the following characters: + - = . _ : /.

See: newTag smart constructor.

Constructors

Tag'

Fields

key :: Text

The tag key (String). The key can't start with aws:.

value :: Text

The value of the tag key.

newTag

Arguments

:: Text
$sel:key:Tag'

-> Text
$sel:value:Tag'

-> Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:Tag', tag_key - The tag key (String). The key can't start with aws:.

$sel:value:Tag', tag_value - The value of the tag key.

tag_key :: Lens' Tag Text

The tag key (String). The key can't start with aws:.

tag_value :: Lens' Tag Text

The value of the tag key.