Network.AWS.EFS.Types.FileSystemSize

Description

Documentation

data FileSystemSize

The latest known metered size (in bytes) of data stored in the file system, in its Value field, and the time at which that size was determined in its Timestamp field. The value doesn't represent the size of a consistent snapshot of the file system, but it is eventually consistent when there are no writes to the file system. That is, the value represents the actual size only if the file system is not modified for a period longer than a couple of hours. Otherwise, the value is not necessarily the exact size the file system was at any instant in time.

See: newFileSystemSize smart constructor.

Constructors

FileSystemSize'

Fields

valueInIA :: Maybe Natural

The latest known metered size (in bytes) of data stored in the Infrequent Access storage class.

valueInStandard :: Maybe Natural

The latest known metered size (in bytes) of data stored in the Standard storage class.

timestamp :: Maybe POSIX

The time at which the size of data, returned in the Value field, was determined. The value is the integer number of seconds since 1970-01-01T00:00:00Z.

value :: Natural

The latest known metered size (in bytes) of data stored in the file system.

newFileSystemSize

Arguments

:: Natural
$sel:value:FileSystemSize'

-> FileSystemSize

Create a value of FileSystemSize with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:valueInIA:FileSystemSize', fileSystemSize_valueInIA - The latest known metered size (in bytes) of data stored in the Infrequent Access storage class.

$sel:valueInStandard:FileSystemSize', fileSystemSize_valueInStandard - The latest known metered size (in bytes) of data stored in the Standard storage class.

$sel:timestamp:FileSystemSize', fileSystemSize_timestamp - The time at which the size of data, returned in the Value field, was determined. The value is the integer number of seconds since 1970-01-01T00:00:00Z.

$sel:value:FileSystemSize', fileSystemSize_value - The latest known metered size (in bytes) of data stored in the file system.

fileSystemSize_valueInIA :: Lens' FileSystemSize (Maybe Natural)

The latest known metered size (in bytes) of data stored in the Infrequent Access storage class.

fileSystemSize_valueInStandard :: Lens' FileSystemSize (Maybe Natural)

The latest known metered size (in bytes) of data stored in the Standard storage class.

fileSystemSize_timestamp :: Lens' FileSystemSize (Maybe UTCTime)

The time at which the size of data, returned in the Value field, was determined. The value is the integer number of seconds since 1970-01-01T00:00:00Z.

fileSystemSize_value :: Lens' FileSystemSize Natural

The latest known metered size (in bytes) of data stored in the file system.