Network.AWS.EFS.Types.LifecyclePolicy

Description

Documentation

data LifecyclePolicy

Describes a policy used by EFS lifecycle management to transition files to the Infrequent Access (IA) storage class.

See: newLifecyclePolicy smart constructor.

Constructors

LifecyclePolicy'

Fields

transitionToIA :: Maybe TransitionToIARules

Describes the period of time that a file is not accessed, after which it transitions to the IA storage class. Metadata operations such as listing the contents of a directory don't count as file access events.

transitionToPrimaryStorageClass :: Maybe TransitionToPrimaryStorageClassRules

Describes the policy used to transition a file from infequent access storage to primary storage.

newLifecyclePolicy :: LifecyclePolicy

Create a value of LifecyclePolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:transitionToIA:LifecyclePolicy', lifecyclePolicy_transitionToIA - Describes the period of time that a file is not accessed, after which it transitions to the IA storage class. Metadata operations such as listing the contents of a directory don't count as file access events.

$sel:transitionToPrimaryStorageClass:LifecyclePolicy', lifecyclePolicy_transitionToPrimaryStorageClass - Describes the policy used to transition a file from infequent access storage to primary storage.

lifecyclePolicy_transitionToIA :: Lens' LifecyclePolicy (Maybe TransitionToIARules)

Describes the period of time that a file is not accessed, after which it transitions to the IA storage class. Metadata operations such as listing the contents of a directory don't count as file access events.

lifecyclePolicy_transitionToPrimaryStorageClass :: Lens' LifecyclePolicy (Maybe TransitionToPrimaryStorageClassRules)

Describes the policy used to transition a file from infequent access storage to primary storage.