Network.AWS.EFS.Types.MountTargetDescription

Description

Documentation

data MountTargetDescription

Provides a description of a mount target.

See: newMountTargetDescription smart constructor.

Constructors

MountTargetDescription'

Fields

ipAddress :: Maybe Text

Address at which the file system can be mounted by using the mount target.

availabilityZoneId :: Maybe Text

The unique and consistent identifier of the Availability Zone that the mount target resides in. For example, use1-az1 is an AZ ID for the us-east-1 Region and it has the same location in every Amazon Web Services account.

vpcId :: Maybe Text

The virtual private cloud (VPC) ID that the mount target is configured in.

availabilityZoneName :: Maybe Text

The name of the Availability Zone in which the mount target is located. Availability Zones are independently mapped to names for each Amazon Web Services account. For example, the Availability Zone us-east-1a for your Amazon Web Services account might not be the same location as us-east-1a for another Amazon Web Services account.

networkInterfaceId :: Maybe Text

The ID of the network interface that Amazon EFS created when it created the mount target.

ownerId :: Maybe Text

Amazon Web Services account ID that owns the resource.

mountTargetId :: Text

System-assigned mount target ID.

fileSystemId :: Text

The ID of the file system for which the mount target is intended.

subnetId :: Text

The ID of the mount target's subnet.

lifeCycleState :: LifeCycleState

Lifecycle state of the mount target.

newMountTargetDescription

Arguments

:: Text
$sel:mountTargetId:MountTargetDescription'

-> Text
$sel:fileSystemId:MountTargetDescription'

-> Text
$sel:subnetId:MountTargetDescription'

-> LifeCycleState
$sel:lifeCycleState:MountTargetDescription'

-> MountTargetDescription

Create a value of MountTargetDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ipAddress:MountTargetDescription', mountTargetDescription_ipAddress - Address at which the file system can be mounted by using the mount target.

$sel:availabilityZoneId:MountTargetDescription', mountTargetDescription_availabilityZoneId - The unique and consistent identifier of the Availability Zone that the mount target resides in. For example, use1-az1 is an AZ ID for the us-east-1 Region and it has the same location in every Amazon Web Services account.

$sel:vpcId:MountTargetDescription', mountTargetDescription_vpcId - The virtual private cloud (VPC) ID that the mount target is configured in.

$sel:availabilityZoneName:MountTargetDescription', mountTargetDescription_availabilityZoneName - The name of the Availability Zone in which the mount target is located. Availability Zones are independently mapped to names for each Amazon Web Services account. For example, the Availability Zone us-east-1a for your Amazon Web Services account might not be the same location as us-east-1a for another Amazon Web Services account.

$sel:networkInterfaceId:MountTargetDescription', mountTargetDescription_networkInterfaceId - The ID of the network interface that Amazon EFS created when it created the mount target.

$sel:ownerId:MountTargetDescription', mountTargetDescription_ownerId - Amazon Web Services account ID that owns the resource.

$sel:mountTargetId:MountTargetDescription', mountTargetDescription_mountTargetId - System-assigned mount target ID.

$sel:fileSystemId:MountTargetDescription', mountTargetDescription_fileSystemId - The ID of the file system for which the mount target is intended.

$sel:subnetId:MountTargetDescription', mountTargetDescription_subnetId - The ID of the mount target's subnet.

$sel:lifeCycleState:MountTargetDescription', mountTargetDescription_lifeCycleState - Lifecycle state of the mount target.

mountTargetDescription_ipAddress :: Lens' MountTargetDescription (Maybe Text)

Address at which the file system can be mounted by using the mount target.

mountTargetDescription_availabilityZoneId :: Lens' MountTargetDescription (Maybe Text)

The unique and consistent identifier of the Availability Zone that the mount target resides in. For example, use1-az1 is an AZ ID for the us-east-1 Region and it has the same location in every Amazon Web Services account.

mountTargetDescription_vpcId :: Lens' MountTargetDescription (Maybe Text)

The virtual private cloud (VPC) ID that the mount target is configured in.

mountTargetDescription_availabilityZoneName :: Lens' MountTargetDescription (Maybe Text)

The name of the Availability Zone in which the mount target is located. Availability Zones are independently mapped to names for each Amazon Web Services account. For example, the Availability Zone us-east-1a for your Amazon Web Services account might not be the same location as us-east-1a for another Amazon Web Services account.

mountTargetDescription_networkInterfaceId :: Lens' MountTargetDescription (Maybe Text)

The ID of the network interface that Amazon EFS created when it created the mount target.

mountTargetDescription_ownerId :: Lens' MountTargetDescription (Maybe Text)

Amazon Web Services account ID that owns the resource.

mountTargetDescription_mountTargetId :: Lens' MountTargetDescription Text

System-assigned mount target ID.

mountTargetDescription_fileSystemId :: Lens' MountTargetDescription Text

The ID of the file system for which the mount target is intended.

mountTargetDescription_subnetId :: Lens' MountTargetDescription Text

The ID of the mount target's subnet.

mountTargetDescription_lifeCycleState :: Lens' MountTargetDescription LifeCycleState

Lifecycle state of the mount target.