Network.AWS.EFS.Types.PosixUser

Description

Documentation

data PosixUser

The full POSIX identity, including the user ID, group ID, and any secondary group IDs, on the access point that is used for all file system operations performed by NFS clients using the access point.

See: newPosixUser smart constructor.

Constructors

PosixUser'

Fields

secondaryGids :: Maybe [Natural]

Secondary POSIX group IDs used for all file system operations using this access point.

uid :: Natural

The POSIX user ID used for all file system operations using this access point.

gid :: Natural

The POSIX group ID used for all file system operations using this access point.

newPosixUser

Arguments

:: Natural
$sel:uid:PosixUser'

-> Natural
$sel:gid:PosixUser'

-> PosixUser

Create a value of PosixUser with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:secondaryGids:PosixUser', posixUser_secondaryGids - Secondary POSIX group IDs used for all file system operations using this access point.

$sel:uid:PosixUser', posixUser_uid - The POSIX user ID used for all file system operations using this access point.

$sel:gid:PosixUser', posixUser_gid - The POSIX group ID used for all file system operations using this access point.

posixUser_secondaryGids :: Lens' PosixUser (Maybe [Natural])

Secondary POSIX group IDs used for all file system operations using this access point.

posixUser_uid :: Lens' PosixUser Natural

The POSIX user ID used for all file system operations using this access point.

posixUser_gid :: Lens' PosixUser Natural

The POSIX group ID used for all file system operations using this access point.