Network.AWS.EFS.Types.ResourceIdPreference

Description

Documentation

data ResourceIdPreference

Describes the resource type and its ID preference for the user's Amazon Web Services account, in the current Amazon Web Services Region.

See: newResourceIdPreference smart constructor.

Constructors

ResourceIdPreference'

Fields

resourceIdType :: Maybe ResourceIdType

Identifies the EFS resource ID preference, either LONG_ID (17 characters) or SHORT_ID (8 characters).

resources :: Maybe [Resource]

Identifies the Amazon EFS resources to which the ID preference setting applies, FILE_SYSTEM and MOUNT_TARGET.

newResourceIdPreference :: ResourceIdPreference

Create a value of ResourceIdPreference with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceIdType:ResourceIdPreference', resourceIdPreference_resourceIdType - Identifies the EFS resource ID preference, either LONG_ID (17 characters) or SHORT_ID (8 characters).

$sel:resources:ResourceIdPreference', resourceIdPreference_resources - Identifies the Amazon EFS resources to which the ID preference setting applies, FILE_SYSTEM and MOUNT_TARGET.

resourceIdPreference_resourceIdType :: Lens' ResourceIdPreference (Maybe ResourceIdType)

Identifies the EFS resource ID preference, either LONG_ID (17 characters) or SHORT_ID (8 characters).

resourceIdPreference_resources :: Lens' ResourceIdPreference (Maybe [Resource])

Identifies the Amazon EFS resources to which the ID preference setting applies, FILE_SYSTEM and MOUNT_TARGET.