Network.AWS.EFS.UpdateFileSystem

Description

Updates the throughput mode or the amount of provisioned throughput of an existing file system.

Creating a Request

data UpdateFileSystem

See: newUpdateFileSystem smart constructor.

Constructors

UpdateFileSystem'

Fields

throughputMode :: Maybe ThroughputMode

(Optional) Updates the file system's throughput mode. If you're not updating your throughput mode, you don't need to provide this value in your request. If you are changing the ThroughputMode to provisioned, you must also set a value for ProvisionedThroughputInMibps.

provisionedThroughputInMibps :: Maybe Double

(Optional) Sets the amount of provisioned throughput, in MiB/s, for the file system. Valid values are 1-1024. If you are changing the throughput mode to provisioned, you must also provide the amount of provisioned throughput. Required if ThroughputMode is changed to provisioned on update.

fileSystemId :: Text

The ID of the file system that you want to update.

newUpdateFileSystem

Arguments

:: Text
$sel:fileSystemId:UpdateFileSystem'

-> UpdateFileSystem

Create a value of UpdateFileSystem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:throughputMode:UpdateFileSystem', updateFileSystem_throughputMode - (Optional) Updates the file system's throughput mode. If you're not updating your throughput mode, you don't need to provide this value in your request. If you are changing the ThroughputMode to provisioned, you must also set a value for ProvisionedThroughputInMibps.

$sel:provisionedThroughputInMibps:UpdateFileSystem', updateFileSystem_provisionedThroughputInMibps - (Optional) Sets the amount of provisioned throughput, in MiB/s, for the file system. Valid values are 1-1024. If you are changing the throughput mode to provisioned, you must also provide the amount of provisioned throughput. Required if ThroughputMode is changed to provisioned on update.

$sel:fileSystemId:UpdateFileSystem', updateFileSystem_fileSystemId - The ID of the file system that you want to update.

Request Lenses

updateFileSystem_throughputMode :: Lens' UpdateFileSystem (Maybe ThroughputMode)

(Optional) Updates the file system's throughput mode. If you're not updating your throughput mode, you don't need to provide this value in your request. If you are changing the ThroughputMode to provisioned, you must also set a value for ProvisionedThroughputInMibps.

updateFileSystem_provisionedThroughputInMibps :: Lens' UpdateFileSystem (Maybe Double)

(Optional) Sets the amount of provisioned throughput, in MiB/s, for the file system. Valid values are 1-1024. If you are changing the throughput mode to provisioned, you must also provide the amount of provisioned throughput. Required if ThroughputMode is changed to provisioned on update.

updateFileSystem_fileSystemId :: Lens' UpdateFileSystem Text

The ID of the file system that you want to update.

Destructuring the Response

data FileSystemDescription

A description of the file system.

See: newFileSystemDescription smart constructor.

Constructors

FileSystemDescription'

Fields

throughputMode :: Maybe ThroughputMode

Displays the file system's throughput mode. For more information, see Throughput modes in the Amazon EFS User Guide.

encrypted :: Maybe Bool

A Boolean value that, if true, indicates that the file system is encrypted.

fileSystemArn :: Maybe Text

The Amazon Resource Name (ARN) for the EFS file system, in the format arn:aws:elasticfilesystem:region:account-id:file-system/file-system-id . Example with sample data: arn:aws:elasticfilesystem:us-west-2:1111333322228888:file-system/fs-01234567

provisionedThroughputInMibps :: Maybe Double

The amount of provisioned throughput, measured in MiB/s, for the file system. Valid for file systems using ThroughputMode set to provisioned.

availabilityZoneName :: Maybe Text

Describes the Amazon Web Services Availability Zone in which the file system is located, and is valid only for file systems using One Zone storage classes. For more information, see Using EFS storage classes in the Amazon EFS User Guide.

availabilityZoneId :: Maybe Text

The unique and consistent identifier of the Availability Zone in which the file system's One Zone storage classes exist. For example, use1-az1 is an Availability Zone ID for the us-east-1 Amazon Web Services Region, and it has the same location in every Amazon Web Services account.

kmsKeyId :: Maybe Text

The ID of an Key Management Service customer master key (CMK) that was used to protect the encrypted file system.

name :: Maybe Text

You can add tags to a file system, including a Name tag. For more information, see CreateFileSystem. If the file system has a Name tag, Amazon EFS returns the value in this field.

ownerId :: Text

The Amazon Web Services account that created the file system. If the file system was created by an IAM user, the parent account to which the user belongs is the owner.

creationToken :: Text

The opaque string specified in the request.

fileSystemId :: Text

The ID of the file system, assigned by Amazon EFS.

creationTime :: POSIX

The time that the file system was created, in seconds (since 1970-01-01T00:00:00Z).

lifeCycleState :: LifeCycleState

The lifecycle phase of the file system.

numberOfMountTargets :: Natural

The current number of mount targets that the file system has. For more information, see CreateMountTarget.

sizeInBytes :: FileSystemSize

The latest known metered size (in bytes) of data stored in the file system, in its Value field, and the time at which that size was determined in its Timestamp field. The Timestamp value is the integer number of seconds since 1970-01-01T00:00:00Z. The SizeInBytes value doesn't represent the size of a consistent snapshot of the file system, but it is eventually consistent when there are no writes to the file system. That is, SizeInBytes represents actual size only if the file system is not modified for a period longer than a couple of hours. Otherwise, the value is not the exact size that the file system was at any point in time.

performanceMode :: PerformanceMode

The performance mode of the file system.

tags :: [Tag]

The tags associated with the file system, presented as an array of Tag objects.

newFileSystemDescription

Arguments

:: Text
$sel:ownerId:FileSystemDescription'

-> Text
$sel:creationToken:FileSystemDescription'

-> Text
$sel:fileSystemId:FileSystemDescription'

-> UTCTime
$sel:creationTime:FileSystemDescription'

-> LifeCycleState
$sel:lifeCycleState:FileSystemDescription'

-> Natural
$sel:numberOfMountTargets:FileSystemDescription'

-> FileSystemSize
$sel:sizeInBytes:FileSystemDescription'

-> PerformanceMode
$sel:performanceMode:FileSystemDescription'

-> FileSystemDescription

Create a value of FileSystemDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:throughputMode:FileSystemDescription', fileSystemDescription_throughputMode - Displays the file system's throughput mode. For more information, see Throughput modes in the Amazon EFS User Guide.

$sel:encrypted:FileSystemDescription', fileSystemDescription_encrypted - A Boolean value that, if true, indicates that the file system is encrypted.

$sel:fileSystemArn:FileSystemDescription', fileSystemDescription_fileSystemArn - The Amazon Resource Name (ARN) for the EFS file system, in the format arn:aws:elasticfilesystem:region:account-id:file-system/file-system-id . Example with sample data: arn:aws:elasticfilesystem:us-west-2:1111333322228888:file-system/fs-01234567

$sel:provisionedThroughputInMibps:FileSystemDescription', fileSystemDescription_provisionedThroughputInMibps - The amount of provisioned throughput, measured in MiB/s, for the file system. Valid for file systems using ThroughputMode set to provisioned.

$sel:availabilityZoneName:FileSystemDescription', fileSystemDescription_availabilityZoneName - Describes the Amazon Web Services Availability Zone in which the file system is located, and is valid only for file systems using One Zone storage classes. For more information, see Using EFS storage classes in the Amazon EFS User Guide.

$sel:availabilityZoneId:FileSystemDescription', fileSystemDescription_availabilityZoneId - The unique and consistent identifier of the Availability Zone in which the file system's One Zone storage classes exist. For example, use1-az1 is an Availability Zone ID for the us-east-1 Amazon Web Services Region, and it has the same location in every Amazon Web Services account.

$sel:kmsKeyId:FileSystemDescription', fileSystemDescription_kmsKeyId - The ID of an Key Management Service customer master key (CMK) that was used to protect the encrypted file system.

$sel:name:FileSystemDescription', fileSystemDescription_name - You can add tags to a file system, including a Name tag. For more information, see CreateFileSystem. If the file system has a Name tag, Amazon EFS returns the value in this field.

$sel:ownerId:FileSystemDescription', fileSystemDescription_ownerId - The Amazon Web Services account that created the file system. If the file system was created by an IAM user, the parent account to which the user belongs is the owner.

$sel:creationToken:FileSystemDescription', fileSystemDescription_creationToken - The opaque string specified in the request.

$sel:fileSystemId:FileSystemDescription', fileSystemDescription_fileSystemId - The ID of the file system, assigned by Amazon EFS.

$sel:creationTime:FileSystemDescription', fileSystemDescription_creationTime - The time that the file system was created, in seconds (since 1970-01-01T00:00:00Z).

$sel:lifeCycleState:FileSystemDescription', fileSystemDescription_lifeCycleState - The lifecycle phase of the file system.

$sel:numberOfMountTargets:FileSystemDescription', fileSystemDescription_numberOfMountTargets - The current number of mount targets that the file system has. For more information, see CreateMountTarget.

$sel:sizeInBytes:FileSystemDescription', fileSystemDescription_sizeInBytes - The latest known metered size (in bytes) of data stored in the file system, in its Value field, and the time at which that size was determined in its Timestamp field. The Timestamp value is the integer number of seconds since 1970-01-01T00:00:00Z. The SizeInBytes value doesn't represent the size of a consistent snapshot of the file system, but it is eventually consistent when there are no writes to the file system. That is, SizeInBytes represents actual size only if the file system is not modified for a period longer than a couple of hours. Otherwise, the value is not the exact size that the file system was at any point in time.

$sel:performanceMode:FileSystemDescription', fileSystemDescription_performanceMode - The performance mode of the file system.

$sel:tags:FileSystemDescription', fileSystemDescription_tags - The tags associated with the file system, presented as an array of Tag objects.

Response Lenses

fileSystemDescription_throughputMode :: Lens' FileSystemDescription (Maybe ThroughputMode)

Displays the file system's throughput mode. For more information, see Throughput modes in the Amazon EFS User Guide.

fileSystemDescription_encrypted :: Lens' FileSystemDescription (Maybe Bool)

A Boolean value that, if true, indicates that the file system is encrypted.

fileSystemDescription_fileSystemArn :: Lens' FileSystemDescription (Maybe Text)

The Amazon Resource Name (ARN) for the EFS file system, in the format arn:aws:elasticfilesystem:region:account-id:file-system/file-system-id . Example with sample data: arn:aws:elasticfilesystem:us-west-2:1111333322228888:file-system/fs-01234567

fileSystemDescription_provisionedThroughputInMibps :: Lens' FileSystemDescription (Maybe Double)

The amount of provisioned throughput, measured in MiB/s, for the file system. Valid for file systems using ThroughputMode set to provisioned.

fileSystemDescription_availabilityZoneName :: Lens' FileSystemDescription (Maybe Text)

Describes the Amazon Web Services Availability Zone in which the file system is located, and is valid only for file systems using One Zone storage classes. For more information, see Using EFS storage classes in the Amazon EFS User Guide.

fileSystemDescription_availabilityZoneId :: Lens' FileSystemDescription (Maybe Text)

The unique and consistent identifier of the Availability Zone in which the file system's One Zone storage classes exist. For example, use1-az1 is an Availability Zone ID for the us-east-1 Amazon Web Services Region, and it has the same location in every Amazon Web Services account.

fileSystemDescription_kmsKeyId :: Lens' FileSystemDescription (Maybe Text)

The ID of an Key Management Service customer master key (CMK) that was used to protect the encrypted file system.

fileSystemDescription_name :: Lens' FileSystemDescription (Maybe Text)

You can add tags to a file system, including a Name tag. For more information, see CreateFileSystem. If the file system has a Name tag, Amazon EFS returns the value in this field.

fileSystemDescription_ownerId :: Lens' FileSystemDescription Text

The Amazon Web Services account that created the file system. If the file system was created by an IAM user, the parent account to which the user belongs is the owner.

fileSystemDescription_creationToken :: Lens' FileSystemDescription Text

The opaque string specified in the request.

fileSystemDescription_fileSystemId :: Lens' FileSystemDescription Text

The ID of the file system, assigned by Amazon EFS.

fileSystemDescription_creationTime :: Lens' FileSystemDescription UTCTime

The time that the file system was created, in seconds (since 1970-01-01T00:00:00Z).

fileSystemDescription_lifeCycleState :: Lens' FileSystemDescription LifeCycleState

The lifecycle phase of the file system.

fileSystemDescription_numberOfMountTargets :: Lens' FileSystemDescription Natural

The current number of mount targets that the file system has. For more information, see CreateMountTarget.

fileSystemDescription_sizeInBytes :: Lens' FileSystemDescription FileSystemSize

The latest known metered size (in bytes) of data stored in the file system, in its Value field, and the time at which that size was determined in its Timestamp field. The Timestamp value is the integer number of seconds since 1970-01-01T00:00:00Z. The SizeInBytes value doesn't represent the size of a consistent snapshot of the file system, but it is eventually consistent when there are no writes to the file system. That is, SizeInBytes represents actual size only if the file system is not modified for a period longer than a couple of hours. Otherwise, the value is not the exact size that the file system was at any point in time.

fileSystemDescription_performanceMode :: Lens' FileSystemDescription PerformanceMode

The performance mode of the file system.

fileSystemDescription_tags :: Lens' FileSystemDescription [Tag]

The tags associated with the file system, presented as an array of Tag objects.