Network.AWS.EKS.CreateAddon

Description

Creates an Amazon EKS add-on.

Amazon EKS add-ons help to automate the provisioning and lifecycle management of common operational software for Amazon EKS clusters. Amazon EKS add-ons can only be used with Amazon EKS clusters running version 1.18 with platform version eks.3 or later because add-ons rely on the Server-side Apply Kubernetes feature, which is only available in Kubernetes 1.18 and later.

Creating a Request

data CreateAddon

See: newCreateAddon smart constructor.

Constructors

CreateAddon'

Fields

addonVersion :: Maybe Text

The version of the add-on. The version must match one of the versions returned by DescribeAddonVersions .

serviceAccountRoleArn :: Maybe Text

The Amazon Resource Name (ARN) of an existing IAM role to bind to the add-on's service account. The role must be assigned the IAM permissions required by the add-on. If you don't specify an existing IAM role, then the add-on uses the permissions assigned to the node IAM role. For more information, see Amazon EKS node IAM role in the Amazon EKS User Guide.

To specify an existing IAM role, you must have an IAM OpenID Connect (OIDC) provider created for your cluster. For more information, see Enabling IAM roles for service accounts on your cluster in the Amazon EKS User Guide.

resolveConflicts :: Maybe ResolveConflicts

How to resolve parameter value conflicts when migrating an existing add-on to an Amazon EKS add-on.

clientRequestToken :: Maybe Text

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

tags :: Maybe (HashMap Text Text)

The metadata to apply to the cluster to assist with categorization and organization. Each tag consists of a key and an optional value, both of which you define.

clusterName :: Text

The name of the cluster to create the add-on for.

addonName :: Text

The name of the add-on. The name must match one of the names returned by DescribeAddonVersions .

newCreateAddon

Arguments

:: Text
$sel:clusterName:CreateAddon'

-> Text
$sel:addonName:CreateAddon'

-> CreateAddon

Create a value of CreateAddon with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:addonVersion:CreateAddon', createAddon_addonVersion - The version of the add-on. The version must match one of the versions returned by DescribeAddonVersions .

$sel:serviceAccountRoleArn:CreateAddon', createAddon_serviceAccountRoleArn - The Amazon Resource Name (ARN) of an existing IAM role to bind to the add-on's service account. The role must be assigned the IAM permissions required by the add-on. If you don't specify an existing IAM role, then the add-on uses the permissions assigned to the node IAM role. For more information, see Amazon EKS node IAM role in the Amazon EKS User Guide.

To specify an existing IAM role, you must have an IAM OpenID Connect (OIDC) provider created for your cluster. For more information, see Enabling IAM roles for service accounts on your cluster in the Amazon EKS User Guide.

$sel:resolveConflicts:CreateAddon', createAddon_resolveConflicts - How to resolve parameter value conflicts when migrating an existing add-on to an Amazon EKS add-on.

$sel:clientRequestToken:CreateAddon', createAddon_clientRequestToken - A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

$sel:tags:CreateAddon', createAddon_tags - The metadata to apply to the cluster to assist with categorization and organization. Each tag consists of a key and an optional value, both of which you define.

$sel:clusterName:CreateAddon', createAddon_clusterName - The name of the cluster to create the add-on for.

$sel:addonName:CreateAddon', createAddon_addonName - The name of the add-on. The name must match one of the names returned by DescribeAddonVersions .

Request Lenses

createAddon_addonVersion :: Lens' CreateAddon (Maybe Text)

The version of the add-on. The version must match one of the versions returned by DescribeAddonVersions .

createAddon_serviceAccountRoleArn :: Lens' CreateAddon (Maybe Text)

The Amazon Resource Name (ARN) of an existing IAM role to bind to the add-on's service account. The role must be assigned the IAM permissions required by the add-on. If you don't specify an existing IAM role, then the add-on uses the permissions assigned to the node IAM role. For more information, see Amazon EKS node IAM role in the Amazon EKS User Guide.

To specify an existing IAM role, you must have an IAM OpenID Connect (OIDC) provider created for your cluster. For more information, see Enabling IAM roles for service accounts on your cluster in the Amazon EKS User Guide.

createAddon_resolveConflicts :: Lens' CreateAddon (Maybe ResolveConflicts)

How to resolve parameter value conflicts when migrating an existing add-on to an Amazon EKS add-on.

createAddon_clientRequestToken :: Lens' CreateAddon (Maybe Text)

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

createAddon_tags :: Lens' CreateAddon (Maybe (HashMap Text Text))

The metadata to apply to the cluster to assist with categorization and organization. Each tag consists of a key and an optional value, both of which you define.

createAddon_clusterName :: Lens' CreateAddon Text

The name of the cluster to create the add-on for.

createAddon_addonName :: Lens' CreateAddon Text

The name of the add-on. The name must match one of the names returned by DescribeAddonVersions .

Destructuring the Response

data CreateAddonResponse

See: newCreateAddonResponse smart constructor.

Constructors

CreateAddonResponse'

Fields

addon :: Maybe Addon

httpStatus :: Int

The response's http status code.

newCreateAddonResponse

Arguments

:: Int
$sel:httpStatus:CreateAddonResponse'

-> CreateAddonResponse

Create a value of CreateAddonResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:addon:CreateAddonResponse', createAddonResponse_addon - Undocumented member.

$sel:httpStatus:CreateAddonResponse', createAddonResponse_httpStatus - The response's http status code.

Response Lenses

createAddonResponse_addon :: Lens' CreateAddonResponse (Maybe Addon)

Undocumented member.

createAddonResponse_httpStatus :: Lens' CreateAddonResponse Int

The response's http status code.