Network.AWS.EKS.CreateFargateProfile

Description

Creates an Fargate profile for your Amazon EKS cluster. You must have at least one Fargate profile in a cluster to be able to run pods on Fargate.

The Fargate profile allows an administrator to declare which pods run on Fargate and specify which pods run on which Fargate profile. This declaration is done through the profile’s selectors. Each profile can have up to five selectors that contain a namespace and labels. A namespace is required for every selector. The label field consists of multiple optional key-value pairs. Pods that match the selectors are scheduled on Fargate. If a to-be-scheduled pod matches any of the selectors in the Fargate profile, then that pod is run on Fargate.

When you create a Fargate profile, you must specify a pod execution role to use with the pods that are scheduled with the profile. This role is added to the cluster's Kubernetes Role Based Access Control (RBAC) for authorization so that the kubelet that is running on the Fargate infrastructure can register with your Amazon EKS cluster so that it can appear in your cluster as a node. The pod execution role also provides IAM permissions to the Fargate infrastructure to allow read access to Amazon ECR image repositories. For more information, see Pod Execution Role in the Amazon EKS User Guide.

Fargate profiles are immutable. However, you can create a new updated profile to replace an existing profile and then delete the original after the updated profile has finished creating.

If any Fargate profiles in a cluster are in the DELETING status, you must wait for that Fargate profile to finish deleting before you can create any other profiles in that cluster.

For more information, see Fargate Profile in the Amazon EKS User Guide.

Creating a Request

data CreateFargateProfile

See: newCreateFargateProfile smart constructor.

Constructors

CreateFargateProfile'

Fields

subnets :: Maybe [Text]

The IDs of subnets to launch your pods into. At this time, pods running on Fargate are not assigned public IP addresses, so only private subnets (with no direct route to an Internet Gateway) are accepted for this parameter.

clientRequestToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

selectors :: Maybe [FargateProfileSelector]

The selectors to match for pods to use this Fargate profile. Each selector must have an associated namespace. Optionally, you can also specify labels for a namespace. You may specify up to five selectors in a Fargate profile.

tags :: Maybe (HashMap Text Text)

The metadata to apply to the Fargate profile to assist with categorization and organization. Each tag consists of a key and an optional value, both of which you define. Fargate profile tags do not propagate to any other resources associated with the Fargate profile, such as the pods that are scheduled with it.

fargateProfileName :: Text

The name of the Fargate profile.

clusterName :: Text

The name of the Amazon EKS cluster to apply the Fargate profile to.

podExecutionRoleArn :: Text

The Amazon Resource Name (ARN) of the pod execution role to use for pods that match the selectors in the Fargate profile. The pod execution role allows Fargate infrastructure to register with your cluster as a node, and it provides read access to Amazon ECR image repositories. For more information, see Pod Execution Role in the Amazon EKS User Guide.

newCreateFargateProfile

Arguments

:: Text
$sel:fargateProfileName:CreateFargateProfile'

-> Text
$sel:clusterName:CreateFargateProfile'

-> Text
$sel:podExecutionRoleArn:CreateFargateProfile'

-> CreateFargateProfile

Create a value of CreateFargateProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:subnets:CreateFargateProfile', createFargateProfile_subnets - The IDs of subnets to launch your pods into. At this time, pods running on Fargate are not assigned public IP addresses, so only private subnets (with no direct route to an Internet Gateway) are accepted for this parameter.

$sel:clientRequestToken:CreateFargateProfile', createFargateProfile_clientRequestToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

$sel:selectors:CreateFargateProfile', createFargateProfile_selectors - The selectors to match for pods to use this Fargate profile. Each selector must have an associated namespace. Optionally, you can also specify labels for a namespace. You may specify up to five selectors in a Fargate profile.

$sel:tags:CreateFargateProfile', createFargateProfile_tags - The metadata to apply to the Fargate profile to assist with categorization and organization. Each tag consists of a key and an optional value, both of which you define. Fargate profile tags do not propagate to any other resources associated with the Fargate profile, such as the pods that are scheduled with it.

$sel:fargateProfileName:CreateFargateProfile', createFargateProfile_fargateProfileName - The name of the Fargate profile.

$sel:clusterName:CreateFargateProfile', createFargateProfile_clusterName - The name of the Amazon EKS cluster to apply the Fargate profile to.

$sel:podExecutionRoleArn:CreateFargateProfile', createFargateProfile_podExecutionRoleArn - The Amazon Resource Name (ARN) of the pod execution role to use for pods that match the selectors in the Fargate profile. The pod execution role allows Fargate infrastructure to register with your cluster as a node, and it provides read access to Amazon ECR image repositories. For more information, see Pod Execution Role in the Amazon EKS User Guide.

Request Lenses

createFargateProfile_subnets :: Lens' CreateFargateProfile (Maybe [Text])

The IDs of subnets to launch your pods into. At this time, pods running on Fargate are not assigned public IP addresses, so only private subnets (with no direct route to an Internet Gateway) are accepted for this parameter.

createFargateProfile_clientRequestToken :: Lens' CreateFargateProfile (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

createFargateProfile_selectors :: Lens' CreateFargateProfile (Maybe [FargateProfileSelector])

The selectors to match for pods to use this Fargate profile. Each selector must have an associated namespace. Optionally, you can also specify labels for a namespace. You may specify up to five selectors in a Fargate profile.

createFargateProfile_tags :: Lens' CreateFargateProfile (Maybe (HashMap Text Text))

The metadata to apply to the Fargate profile to assist with categorization and organization. Each tag consists of a key and an optional value, both of which you define. Fargate profile tags do not propagate to any other resources associated with the Fargate profile, such as the pods that are scheduled with it.

createFargateProfile_fargateProfileName :: Lens' CreateFargateProfile Text

The name of the Fargate profile.

createFargateProfile_clusterName :: Lens' CreateFargateProfile Text

The name of the Amazon EKS cluster to apply the Fargate profile to.

createFargateProfile_podExecutionRoleArn :: Lens' CreateFargateProfile Text

The Amazon Resource Name (ARN) of the pod execution role to use for pods that match the selectors in the Fargate profile. The pod execution role allows Fargate infrastructure to register with your cluster as a node, and it provides read access to Amazon ECR image repositories. For more information, see Pod Execution Role in the Amazon EKS User Guide.

Destructuring the Response

data CreateFargateProfileResponse

See: newCreateFargateProfileResponse smart constructor.

Constructors

CreateFargateProfileResponse'

Fields

fargateProfile :: Maybe FargateProfile

The full description of your new Fargate profile.

httpStatus :: Int

The response's http status code.

newCreateFargateProfileResponse

Arguments

:: Int
$sel:httpStatus:CreateFargateProfileResponse'

-> CreateFargateProfileResponse

Create a value of CreateFargateProfileResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fargateProfile:CreateFargateProfileResponse', createFargateProfileResponse_fargateProfile - The full description of your new Fargate profile.

$sel:httpStatus:CreateFargateProfileResponse', createFargateProfileResponse_httpStatus - The response's http status code.

Response Lenses

createFargateProfileResponse_fargateProfile :: Lens' CreateFargateProfileResponse (Maybe FargateProfile)

The full description of your new Fargate profile.

createFargateProfileResponse_httpStatus :: Lens' CreateFargateProfileResponse Int

The response's http status code.