Network.AWS.EKS.CreateNodegroup

Description

Creates a managed node group for an Amazon EKS cluster. You can only create a node group for your cluster that is equal to the current Kubernetes version for the cluster. All node groups are created with the latest AMI release version for the respective minor Kubernetes version of the cluster, unless you deploy a custom AMI using a launch template. For more information about using launch templates, see Launch template support.

An Amazon EKS managed node group is an Amazon EC2 Auto Scaling group and associated Amazon EC2 instances that are managed by Amazon Web Services for an Amazon EKS cluster. Each node group uses a version of the Amazon EKS optimized Amazon Linux 2 AMI. For more information, see Managed Node Groups in the Amazon EKS User Guide.

Creating a Request

data CreateNodegroup

See: newCreateNodegroup smart constructor.

Constructors

CreateNodegroup'

Fields

scalingConfig :: Maybe NodegroupScalingConfig

The scaling configuration details for the Auto Scaling group that is created for your node group.

capacityType :: Maybe CapacityTypes

The capacity type for your node group.

releaseVersion :: Maybe Text

The AMI version of the Amazon EKS optimized AMI to use with your node group. By default, the latest available AMI version for the node group's current Kubernetes version is used. For more information, see Amazon EKS optimized Amazon Linux 2 AMI versions in the Amazon EKS User Guide. If you specify launchTemplate, and your launch template uses a custom AMI, then don't specify releaseVersion, or the node group deployment will fail. For more information about using launch templates with Amazon EKS, see Launch template support in the Amazon EKS User Guide.

remoteAccess :: Maybe RemoteAccessConfig

The remote access (SSH) configuration to use with your node group. If you specify launchTemplate, then don't specify remoteAccess, or the node group deployment will fail. For more information about using launch templates with Amazon EKS, see Launch template support in the Amazon EKS User Guide.

diskSize :: Maybe Int

The root device disk size (in GiB) for your node group instances. The default disk size is 20 GiB. If you specify launchTemplate, then don't specify diskSize, or the node group deployment will fail. For more information about using launch templates with Amazon EKS, see Launch template support in the Amazon EKS User Guide.

labels :: Maybe (HashMap Text Text)

The Kubernetes labels to be applied to the nodes in the node group when they are created.

launchTemplate :: Maybe LaunchTemplateSpecification

An object representing a node group's launch template specification. If specified, then do not specify instanceTypes, diskSize, or remoteAccess and make sure that the launch template meets the requirements in launchTemplateSpecification.

version :: Maybe Text

The Kubernetes version to use for your managed nodes. By default, the Kubernetes version of the cluster is used, and this is the only accepted specified value. If you specify launchTemplate, and your launch template uses a custom AMI, then don't specify version, or the node group deployment will fail. For more information about using launch templates with Amazon EKS, see Launch template support in the Amazon EKS User Guide.

tags :: Maybe (HashMap Text Text)

The metadata to apply to the node group to assist with categorization and organization. Each tag consists of a key and an optional value, both of which you define. Node group tags do not propagate to any other resources associated with the node group, such as the Amazon EC2 instances or subnets.

updateConfig :: Maybe NodegroupUpdateConfig

The node group update configuration.

clientRequestToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

amiType :: Maybe AMITypes

The AMI type for your node group. GPU instance types should use the AL2_x86_64_GPU AMI type. Non-GPU instances should use the AL2_x86_64 AMI type. Arm instances should use the AL2_ARM_64 AMI type. All types use the Amazon EKS optimized Amazon Linux 2 AMI. If you specify launchTemplate, and your launch template uses a custom AMI, then don't specify amiType, or the node group deployment will fail. For more information about using launch templates with Amazon EKS, see Launch template support in the Amazon EKS User Guide.

taints :: Maybe [Taint]

The Kubernetes taints to be applied to the nodes in the node group.

instanceTypes :: Maybe [Text]

Specify the instance types for a node group. If you specify a GPU instance type, be sure to specify AL2_x86_64_GPU with the amiType parameter. If you specify launchTemplate, then you can specify zero or one instance type in your launch template or you can specify 0-20 instance types for instanceTypes. If however, you specify an instance type in your launch template and specify any instanceTypes, the node group deployment will fail. If you don't specify an instance type in a launch template or for instanceTypes, then t3.medium is used, by default. If you specify Spot for capacityType, then we recommend specifying multiple values for instanceTypes. For more information, see Managed node group capacity types and Launch template support in the Amazon EKS User Guide.

clusterName :: Text

The name of the cluster to create the node group in.

nodegroupName :: Text

The unique name to give your node group.

subnets :: [Text]

The subnets to use for the Auto Scaling group that is created for your node group. If you specify launchTemplate, then don't specify SubnetId in your launch template, or the node group deployment will fail. For more information about using launch templates with Amazon EKS, see Launch template support in the Amazon EKS User Guide.

nodeRole :: Text

The Amazon Resource Name (ARN) of the IAM role to associate with your node group. The Amazon EKS worker node kubelet daemon makes calls to Amazon Web Services APIs on your behalf. Nodes receive permissions for these API calls through an IAM instance profile and associated policies. Before you can launch nodes and register them into a cluster, you must create an IAM role for those nodes to use when they are launched. For more information, see Amazon EKS node IAM role in the /Amazon EKS User Guide/ . If you specify launchTemplate, then don't specify IamInstanceProfile in your launch template, or the node group deployment will fail. For more information about using launch templates with Amazon EKS, see Launch template support in the Amazon EKS User Guide.

newCreateNodegroup

Arguments

:: Text
$sel:clusterName:CreateNodegroup'

-> Text
$sel:nodegroupName:CreateNodegroup'

-> Text
$sel:nodeRole:CreateNodegroup'

-> CreateNodegroup

Create a value of CreateNodegroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:scalingConfig:CreateNodegroup', createNodegroup_scalingConfig - The scaling configuration details for the Auto Scaling group that is created for your node group.

$sel:capacityType:CreateNodegroup', createNodegroup_capacityType - The capacity type for your node group.

$sel:releaseVersion:CreateNodegroup', createNodegroup_releaseVersion - The AMI version of the Amazon EKS optimized AMI to use with your node group. By default, the latest available AMI version for the node group's current Kubernetes version is used. For more information, see Amazon EKS optimized Amazon Linux 2 AMI versions in the Amazon EKS User Guide. If you specify launchTemplate, and your launch template uses a custom AMI, then don't specify releaseVersion, or the node group deployment will fail. For more information about using launch templates with Amazon EKS, see Launch template support in the Amazon EKS User Guide.

$sel:remoteAccess:CreateNodegroup', createNodegroup_remoteAccess - The remote access (SSH) configuration to use with your node group. If you specify launchTemplate, then don't specify remoteAccess, or the node group deployment will fail. For more information about using launch templates with Amazon EKS, see Launch template support in the Amazon EKS User Guide.

$sel:diskSize:CreateNodegroup', createNodegroup_diskSize - The root device disk size (in GiB) for your node group instances. The default disk size is 20 GiB. If you specify launchTemplate, then don't specify diskSize, or the node group deployment will fail. For more information about using launch templates with Amazon EKS, see Launch template support in the Amazon EKS User Guide.

$sel:labels:CreateNodegroup', createNodegroup_labels - The Kubernetes labels to be applied to the nodes in the node group when they are created.

$sel:launchTemplate:CreateNodegroup', createNodegroup_launchTemplate - An object representing a node group's launch template specification. If specified, then do not specify instanceTypes, diskSize, or remoteAccess and make sure that the launch template meets the requirements in launchTemplateSpecification.

$sel:version:CreateNodegroup', createNodegroup_version - The Kubernetes version to use for your managed nodes. By default, the Kubernetes version of the cluster is used, and this is the only accepted specified value. If you specify launchTemplate, and your launch template uses a custom AMI, then don't specify version, or the node group deployment will fail. For more information about using launch templates with Amazon EKS, see Launch template support in the Amazon EKS User Guide.

$sel:tags:CreateNodegroup', createNodegroup_tags - The metadata to apply to the node group to assist with categorization and organization. Each tag consists of a key and an optional value, both of which you define. Node group tags do not propagate to any other resources associated with the node group, such as the Amazon EC2 instances or subnets.

$sel:updateConfig:CreateNodegroup', createNodegroup_updateConfig - The node group update configuration.

$sel:clientRequestToken:CreateNodegroup', createNodegroup_clientRequestToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

$sel:amiType:CreateNodegroup', createNodegroup_amiType - The AMI type for your node group. GPU instance types should use the AL2_x86_64_GPU AMI type. Non-GPU instances should use the AL2_x86_64 AMI type. Arm instances should use the AL2_ARM_64 AMI type. All types use the Amazon EKS optimized Amazon Linux 2 AMI. If you specify launchTemplate, and your launch template uses a custom AMI, then don't specify amiType, or the node group deployment will fail. For more information about using launch templates with Amazon EKS, see Launch template support in the Amazon EKS User Guide.

$sel:taints:CreateNodegroup', createNodegroup_taints - The Kubernetes taints to be applied to the nodes in the node group.

$sel:instanceTypes:CreateNodegroup', createNodegroup_instanceTypes - Specify the instance types for a node group. If you specify a GPU instance type, be sure to specify AL2_x86_64_GPU with the amiType parameter. If you specify launchTemplate, then you can specify zero or one instance type in your launch template or you can specify 0-20 instance types for instanceTypes. If however, you specify an instance type in your launch template and specify any instanceTypes, the node group deployment will fail. If you don't specify an instance type in a launch template or for instanceTypes, then t3.medium is used, by default. If you specify Spot for capacityType, then we recommend specifying multiple values for instanceTypes. For more information, see Managed node group capacity types and Launch template support in the Amazon EKS User Guide.

$sel:clusterName:CreateNodegroup', createNodegroup_clusterName - The name of the cluster to create the node group in.

$sel:nodegroupName:CreateNodegroup', createNodegroup_nodegroupName - The unique name to give your node group.

$sel:subnets:CreateNodegroup', createNodegroup_subnets - The subnets to use for the Auto Scaling group that is created for your node group. If you specify launchTemplate, then don't specify SubnetId in your launch template, or the node group deployment will fail. For more information about using launch templates with Amazon EKS, see Launch template support in the Amazon EKS User Guide.

$sel:nodeRole:CreateNodegroup', createNodegroup_nodeRole - The Amazon Resource Name (ARN) of the IAM role to associate with your node group. The Amazon EKS worker node kubelet daemon makes calls to Amazon Web Services APIs on your behalf. Nodes receive permissions for these API calls through an IAM instance profile and associated policies. Before you can launch nodes and register them into a cluster, you must create an IAM role for those nodes to use when they are launched. For more information, see Amazon EKS node IAM role in the /Amazon EKS User Guide/ . If you specify launchTemplate, then don't specify IamInstanceProfile in your launch template, or the node group deployment will fail. For more information about using launch templates with Amazon EKS, see Launch template support in the Amazon EKS User Guide.

Request Lenses

createNodegroup_scalingConfig :: Lens' CreateNodegroup (Maybe NodegroupScalingConfig)

The scaling configuration details for the Auto Scaling group that is created for your node group.

createNodegroup_capacityType :: Lens' CreateNodegroup (Maybe CapacityTypes)

The capacity type for your node group.

createNodegroup_releaseVersion :: Lens' CreateNodegroup (Maybe Text)

The AMI version of the Amazon EKS optimized AMI to use with your node group. By default, the latest available AMI version for the node group's current Kubernetes version is used. For more information, see Amazon EKS optimized Amazon Linux 2 AMI versions in the Amazon EKS User Guide. If you specify launchTemplate, and your launch template uses a custom AMI, then don't specify releaseVersion, or the node group deployment will fail. For more information about using launch templates with Amazon EKS, see Launch template support in the Amazon EKS User Guide.

createNodegroup_remoteAccess :: Lens' CreateNodegroup (Maybe RemoteAccessConfig)

The remote access (SSH) configuration to use with your node group. If you specify launchTemplate, then don't specify remoteAccess, or the node group deployment will fail. For more information about using launch templates with Amazon EKS, see Launch template support in the Amazon EKS User Guide.

createNodegroup_diskSize :: Lens' CreateNodegroup (Maybe Int)

The root device disk size (in GiB) for your node group instances. The default disk size is 20 GiB. If you specify launchTemplate, then don't specify diskSize, or the node group deployment will fail. For more information about using launch templates with Amazon EKS, see Launch template support in the Amazon EKS User Guide.

createNodegroup_labels :: Lens' CreateNodegroup (Maybe (HashMap Text Text))

The Kubernetes labels to be applied to the nodes in the node group when they are created.

createNodegroup_launchTemplate :: Lens' CreateNodegroup (Maybe LaunchTemplateSpecification)

An object representing a node group's launch template specification. If specified, then do not specify instanceTypes, diskSize, or remoteAccess and make sure that the launch template meets the requirements in launchTemplateSpecification.

createNodegroup_version :: Lens' CreateNodegroup (Maybe Text)

The Kubernetes version to use for your managed nodes. By default, the Kubernetes version of the cluster is used, and this is the only accepted specified value. If you specify launchTemplate, and your launch template uses a custom AMI, then don't specify version, or the node group deployment will fail. For more information about using launch templates with Amazon EKS, see Launch template support in the Amazon EKS User Guide.

createNodegroup_tags :: Lens' CreateNodegroup (Maybe (HashMap Text Text))

The metadata to apply to the node group to assist with categorization and organization. Each tag consists of a key and an optional value, both of which you define. Node group tags do not propagate to any other resources associated with the node group, such as the Amazon EC2 instances or subnets.

createNodegroup_updateConfig :: Lens' CreateNodegroup (Maybe NodegroupUpdateConfig)

The node group update configuration.

createNodegroup_clientRequestToken :: Lens' CreateNodegroup (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

createNodegroup_amiType :: Lens' CreateNodegroup (Maybe AMITypes)

The AMI type for your node group. GPU instance types should use the AL2_x86_64_GPU AMI type. Non-GPU instances should use the AL2_x86_64 AMI type. Arm instances should use the AL2_ARM_64 AMI type. All types use the Amazon EKS optimized Amazon Linux 2 AMI. If you specify launchTemplate, and your launch template uses a custom AMI, then don't specify amiType, or the node group deployment will fail. For more information about using launch templates with Amazon EKS, see Launch template support in the Amazon EKS User Guide.

createNodegroup_taints :: Lens' CreateNodegroup (Maybe [Taint])

The Kubernetes taints to be applied to the nodes in the node group.

createNodegroup_instanceTypes :: Lens' CreateNodegroup (Maybe [Text])

Specify the instance types for a node group. If you specify a GPU instance type, be sure to specify AL2_x86_64_GPU with the amiType parameter. If you specify launchTemplate, then you can specify zero or one instance type in your launch template or you can specify 0-20 instance types for instanceTypes. If however, you specify an instance type in your launch template and specify any instanceTypes, the node group deployment will fail. If you don't specify an instance type in a launch template or for instanceTypes, then t3.medium is used, by default. If you specify Spot for capacityType, then we recommend specifying multiple values for instanceTypes. For more information, see Managed node group capacity types and Launch template support in the Amazon EKS User Guide.

createNodegroup_clusterName :: Lens' CreateNodegroup Text

The name of the cluster to create the node group in.

createNodegroup_nodegroupName :: Lens' CreateNodegroup Text

The unique name to give your node group.

createNodegroup_subnets :: Lens' CreateNodegroup [Text]

The subnets to use for the Auto Scaling group that is created for your node group. If you specify launchTemplate, then don't specify SubnetId in your launch template, or the node group deployment will fail. For more information about using launch templates with Amazon EKS, see Launch template support in the Amazon EKS User Guide.

createNodegroup_nodeRole :: Lens' CreateNodegroup Text

The Amazon Resource Name (ARN) of the IAM role to associate with your node group. The Amazon EKS worker node kubelet daemon makes calls to Amazon Web Services APIs on your behalf. Nodes receive permissions for these API calls through an IAM instance profile and associated policies. Before you can launch nodes and register them into a cluster, you must create an IAM role for those nodes to use when they are launched. For more information, see Amazon EKS node IAM role in the /Amazon EKS User Guide/ . If you specify launchTemplate, then don't specify IamInstanceProfile in your launch template, or the node group deployment will fail. For more information about using launch templates with Amazon EKS, see Launch template support in the Amazon EKS User Guide.

Destructuring the Response

data CreateNodegroupResponse

See: newCreateNodegroupResponse smart constructor.

Constructors

CreateNodegroupResponse'

Fields

nodegroup :: Maybe Nodegroup

The full description of your new node group.

httpStatus :: Int

The response's http status code.

newCreateNodegroupResponse

Arguments

:: Int
$sel:httpStatus:CreateNodegroupResponse'

-> CreateNodegroupResponse

Create a value of CreateNodegroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nodegroup:CreateNodegroupResponse', createNodegroupResponse_nodegroup - The full description of your new node group.

$sel:httpStatus:CreateNodegroupResponse', createNodegroupResponse_httpStatus - The response's http status code.

Response Lenses

createNodegroupResponse_nodegroup :: Lens' CreateNodegroupResponse (Maybe Nodegroup)

The full description of your new node group.

createNodegroupResponse_httpStatus :: Lens' CreateNodegroupResponse Int

The response's http status code.