Network.AWS.EKS.DeleteNodegroup

Description

Deletes an Amazon EKS node group for a cluster.

Creating a Request

data DeleteNodegroup

See: newDeleteNodegroup smart constructor.

Constructors

DeleteNodegroup'

Fields

clusterName :: Text

The name of the Amazon EKS cluster that is associated with your node group.

nodegroupName :: Text

The name of the node group to delete.

newDeleteNodegroup

Arguments

:: Text
$sel:clusterName:DeleteNodegroup'

-> Text
$sel:nodegroupName:DeleteNodegroup'

-> DeleteNodegroup

Create a value of DeleteNodegroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clusterName:DeleteNodegroup', deleteNodegroup_clusterName - The name of the Amazon EKS cluster that is associated with your node group.

$sel:nodegroupName:DeleteNodegroup', deleteNodegroup_nodegroupName - The name of the node group to delete.

Request Lenses

deleteNodegroup_clusterName :: Lens' DeleteNodegroup Text

The name of the Amazon EKS cluster that is associated with your node group.

deleteNodegroup_nodegroupName :: Lens' DeleteNodegroup Text

The name of the node group to delete.

Destructuring the Response

data DeleteNodegroupResponse

See: newDeleteNodegroupResponse smart constructor.

Constructors

DeleteNodegroupResponse'

Fields

nodegroup :: Maybe Nodegroup

The full description of your deleted node group.

httpStatus :: Int

The response's http status code.

newDeleteNodegroupResponse

Arguments

:: Int
$sel:httpStatus:DeleteNodegroupResponse'

-> DeleteNodegroupResponse

Create a value of DeleteNodegroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nodegroup:DeleteNodegroupResponse', deleteNodegroupResponse_nodegroup - The full description of your deleted node group.

$sel:httpStatus:DeleteNodegroupResponse', deleteNodegroupResponse_httpStatus - The response's http status code.

Response Lenses

deleteNodegroupResponse_nodegroup :: Lens' DeleteNodegroupResponse (Maybe Nodegroup)

The full description of your deleted node group.

deleteNodegroupResponse_httpStatus :: Lens' DeleteNodegroupResponse Int

The response's http status code.