Network.AWS.EKS.DescribeAddonVersions

Description

Describes the Kubernetes versions that the add-on can be used with.

This operation returns paginated results.

Creating a Request

data DescribeAddonVersions

See: newDescribeAddonVersions smart constructor.

Constructors

DescribeAddonVersions'

Fields

nextToken :: Maybe Text

The nextToken value returned from a previous paginated DescribeAddonVersionsRequest where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value.

This token should be treated as an opaque identifier that is used only to retrieve the next items in a list and not for other programmatic purposes.

maxResults :: Maybe Natural

The maximum number of results to return.

kubernetesVersion :: Maybe Text

The Kubernetes versions that the add-on can be used with.

addonName :: Maybe Text

The name of the add-on. The name must match one of the names returned by ListAddons .

newDescribeAddonVersions :: DescribeAddonVersions

Create a value of DescribeAddonVersions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeAddonVersions', describeAddonVersions_nextToken - The nextToken value returned from a previous paginated DescribeAddonVersionsRequest where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value.

This token should be treated as an opaque identifier that is used only to retrieve the next items in a list and not for other programmatic purposes.

$sel:maxResults:DescribeAddonVersions', describeAddonVersions_maxResults - The maximum number of results to return.

$sel:kubernetesVersion:DescribeAddonVersions', describeAddonVersions_kubernetesVersion - The Kubernetes versions that the add-on can be used with.

$sel:addonName:DescribeAddonVersions', describeAddonVersions_addonName - The name of the add-on. The name must match one of the names returned by ListAddons .

Request Lenses

describeAddonVersions_nextToken :: Lens' DescribeAddonVersions (Maybe Text)

The nextToken value returned from a previous paginated DescribeAddonVersionsRequest where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value.

This token should be treated as an opaque identifier that is used only to retrieve the next items in a list and not for other programmatic purposes.

describeAddonVersions_maxResults :: Lens' DescribeAddonVersions (Maybe Natural)

The maximum number of results to return.

describeAddonVersions_kubernetesVersion :: Lens' DescribeAddonVersions (Maybe Text)

The Kubernetes versions that the add-on can be used with.

describeAddonVersions_addonName :: Lens' DescribeAddonVersions (Maybe Text)

The name of the add-on. The name must match one of the names returned by ListAddons .

Destructuring the Response

data DescribeAddonVersionsResponse

See: newDescribeAddonVersionsResponse smart constructor.

Constructors

DescribeAddonVersionsResponse'

Fields

nextToken :: Maybe Text

The nextToken value returned from a previous paginated DescribeAddonVersionsResponse where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value.

This token should be treated as an opaque identifier that is used only to retrieve the next items in a list and not for other programmatic purposes.

addons :: Maybe [AddonInfo]

The list of available versions with Kubernetes version compatibility.

httpStatus :: Int

The response's http status code.

newDescribeAddonVersionsResponse

Arguments

:: Int
$sel:httpStatus:DescribeAddonVersionsResponse'

-> DescribeAddonVersionsResponse

Create a value of DescribeAddonVersionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeAddonVersions', describeAddonVersionsResponse_nextToken - The nextToken value returned from a previous paginated DescribeAddonVersionsResponse where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value.

This token should be treated as an opaque identifier that is used only to retrieve the next items in a list and not for other programmatic purposes.

$sel:addons:DescribeAddonVersionsResponse', describeAddonVersionsResponse_addons - The list of available versions with Kubernetes version compatibility.

$sel:httpStatus:DescribeAddonVersionsResponse', describeAddonVersionsResponse_httpStatus - The response's http status code.

Response Lenses

describeAddonVersionsResponse_nextToken :: Lens' DescribeAddonVersionsResponse (Maybe Text)

The nextToken value returned from a previous paginated DescribeAddonVersionsResponse where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value.

This token should be treated as an opaque identifier that is used only to retrieve the next items in a list and not for other programmatic purposes.

describeAddonVersionsResponse_addons :: Lens' DescribeAddonVersionsResponse (Maybe [AddonInfo])

The list of available versions with Kubernetes version compatibility.

describeAddonVersionsResponse_httpStatus :: Lens' DescribeAddonVersionsResponse Int

The response's http status code.