Network.AWS.EKS.DescribeNodegroup

Description

Returns descriptive information about an Amazon EKS node group.

Creating a Request

data DescribeNodegroup

See: newDescribeNodegroup smart constructor.

Constructors

DescribeNodegroup'

Fields

clusterName :: Text

The name of the Amazon EKS cluster associated with the node group.

nodegroupName :: Text

The name of the node group to describe.

newDescribeNodegroup

Arguments

:: Text
$sel:clusterName:DescribeNodegroup'

-> Text
$sel:nodegroupName:DescribeNodegroup'

-> DescribeNodegroup

Create a value of DescribeNodegroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clusterName:DescribeNodegroup', describeNodegroup_clusterName - The name of the Amazon EKS cluster associated with the node group.

$sel:nodegroupName:DescribeNodegroup', describeNodegroup_nodegroupName - The name of the node group to describe.

Request Lenses

describeNodegroup_clusterName :: Lens' DescribeNodegroup Text

The name of the Amazon EKS cluster associated with the node group.

describeNodegroup_nodegroupName :: Lens' DescribeNodegroup Text

The name of the node group to describe.

Destructuring the Response

data DescribeNodegroupResponse

See: newDescribeNodegroupResponse smart constructor.

Constructors

DescribeNodegroupResponse'

Fields

nodegroup :: Maybe Nodegroup

The full description of your node group.

httpStatus :: Int

The response's http status code.

newDescribeNodegroupResponse

Arguments

:: Int
$sel:httpStatus:DescribeNodegroupResponse'

-> DescribeNodegroupResponse

Create a value of DescribeNodegroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nodegroup:DescribeNodegroupResponse', describeNodegroupResponse_nodegroup - The full description of your node group.

$sel:httpStatus:DescribeNodegroupResponse', describeNodegroupResponse_httpStatus - The response's http status code.

Response Lenses

describeNodegroupResponse_nodegroup :: Lens' DescribeNodegroupResponse (Maybe Nodegroup)

The full description of your node group.

describeNodegroupResponse_httpStatus :: Lens' DescribeNodegroupResponse Int

The response's http status code.