Network.AWS.EKS.DescribeUpdate

Description

Returns descriptive information about an update against your Amazon EKS cluster or associated managed node group.

When the status of the update is Succeeded, the update is complete. If an update fails, the status is Failed, and an error detail explains the reason for the failure.

Creating a Request

data DescribeUpdate

See: newDescribeUpdate smart constructor.

Constructors

DescribeUpdate'

Fields

nodegroupName :: Maybe Text

The name of the Amazon EKS node group associated with the update.

addonName :: Maybe Text

The name of the add-on. The name must match one of the names returned by ListAddons .

name :: Text

The name of the Amazon EKS cluster associated with the update.

updateId :: Text

The ID of the update to describe.

newDescribeUpdate

Arguments

:: Text
$sel:name:DescribeUpdate'

-> Text
$sel:updateId:DescribeUpdate'

-> DescribeUpdate

Create a value of DescribeUpdate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nodegroupName:DescribeUpdate', describeUpdate_nodegroupName - The name of the Amazon EKS node group associated with the update.

$sel:addonName:DescribeUpdate', describeUpdate_addonName - The name of the add-on. The name must match one of the names returned by ListAddons .

$sel:name:DescribeUpdate', describeUpdate_name - The name of the Amazon EKS cluster associated with the update.

$sel:updateId:DescribeUpdate', describeUpdate_updateId - The ID of the update to describe.

Request Lenses

describeUpdate_nodegroupName :: Lens' DescribeUpdate (Maybe Text)

The name of the Amazon EKS node group associated with the update.

describeUpdate_addonName :: Lens' DescribeUpdate (Maybe Text)

The name of the add-on. The name must match one of the names returned by ListAddons .

describeUpdate_name :: Lens' DescribeUpdate Text

The name of the Amazon EKS cluster associated with the update.

describeUpdate_updateId :: Lens' DescribeUpdate Text

The ID of the update to describe.

Destructuring the Response

data DescribeUpdateResponse

See: newDescribeUpdateResponse smart constructor.

Constructors

DescribeUpdateResponse'

Fields

update :: Maybe Update

The full description of the specified update.

httpStatus :: Int

The response's http status code.

newDescribeUpdateResponse

Arguments

:: Int
$sel:httpStatus:DescribeUpdateResponse'

-> DescribeUpdateResponse

Create a value of DescribeUpdateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:update:DescribeUpdateResponse', describeUpdateResponse_update - The full description of the specified update.

$sel:httpStatus:DescribeUpdateResponse', describeUpdateResponse_httpStatus - The response's http status code.

Response Lenses

describeUpdateResponse_update :: Lens' DescribeUpdateResponse (Maybe Update)

The full description of the specified update.

describeUpdateResponse_httpStatus :: Lens' DescribeUpdateResponse Int

The response's http status code.