Network.AWS.EKS.ListClusters

Description

Lists the Amazon EKS clusters in your Amazon Web Services account in the specified Region.

This operation returns paginated results.

Creating a Request

data ListClusters

See: newListClusters smart constructor.

Constructors

ListClusters'

Fields

nextToken :: Maybe Text

The nextToken value returned from a previous paginated ListClusters request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value.

This token should be treated as an opaque identifier that is used only to retrieve the next items in a list and not for other programmatic purposes.

maxResults :: Maybe Natural

The maximum number of cluster results returned by ListClusters in paginated output. When you use this parameter, ListClusters returns only maxResults results in a single page along with a nextToken response element. You can see the remaining results of the initial request by sending another ListClusters request with the returned nextToken value. This value can be between 1 and 100. If you don't use this parameter, ListClusters returns up to 100 results and a nextToken value if applicable.

include :: Maybe [Text]

Indicates whether connected clusters are included in the returned list. Default value is 'ALL'.

newListClusters :: ListClusters

Create a value of ListClusters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListClusters', listClusters_nextToken - The nextToken value returned from a previous paginated ListClusters request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value.

This token should be treated as an opaque identifier that is used only to retrieve the next items in a list and not for other programmatic purposes.

$sel:maxResults:ListClusters', listClusters_maxResults - The maximum number of cluster results returned by ListClusters in paginated output. When you use this parameter, ListClusters returns only maxResults results in a single page along with a nextToken response element. You can see the remaining results of the initial request by sending another ListClusters request with the returned nextToken value. This value can be between 1 and 100. If you don't use this parameter, ListClusters returns up to 100 results and a nextToken value if applicable.

$sel:include:ListClusters', listClusters_include - Indicates whether connected clusters are included in the returned list. Default value is 'ALL'.

Request Lenses

listClusters_nextToken :: Lens' ListClusters (Maybe Text)

The nextToken value returned from a previous paginated ListClusters request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value.

This token should be treated as an opaque identifier that is used only to retrieve the next items in a list and not for other programmatic purposes.

listClusters_maxResults :: Lens' ListClusters (Maybe Natural)

The maximum number of cluster results returned by ListClusters in paginated output. When you use this parameter, ListClusters returns only maxResults results in a single page along with a nextToken response element. You can see the remaining results of the initial request by sending another ListClusters request with the returned nextToken value. This value can be between 1 and 100. If you don't use this parameter, ListClusters returns up to 100 results and a nextToken value if applicable.

listClusters_include :: Lens' ListClusters (Maybe [Text])

Indicates whether connected clusters are included in the returned list. Default value is 'ALL'.

Destructuring the Response

data ListClustersResponse

See: newListClustersResponse smart constructor.

Constructors

ListClustersResponse'

Fields

nextToken :: Maybe Text

The nextToken value to include in a future ListClusters request. When the results of a ListClusters request exceed maxResults, you can use this value to retrieve the next page of results. This value is null when there are no more results to return.

clusters :: Maybe [Text]

A list of all of the clusters for your account in the specified Region.

httpStatus :: Int

The response's http status code.

newListClustersResponse

Arguments

:: Int
$sel:httpStatus:ListClustersResponse'

-> ListClustersResponse

Create a value of ListClustersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListClusters', listClustersResponse_nextToken - The nextToken value to include in a future ListClusters request. When the results of a ListClusters request exceed maxResults, you can use this value to retrieve the next page of results. This value is null when there are no more results to return.

$sel:clusters:ListClustersResponse', listClustersResponse_clusters - A list of all of the clusters for your account in the specified Region.

$sel:httpStatus:ListClustersResponse', listClustersResponse_httpStatus - The response's http status code.

Response Lenses

listClustersResponse_nextToken :: Lens' ListClustersResponse (Maybe Text)

The nextToken value to include in a future ListClusters request. When the results of a ListClusters request exceed maxResults, you can use this value to retrieve the next page of results. This value is null when there are no more results to return.

listClustersResponse_clusters :: Lens' ListClustersResponse (Maybe [Text])

A list of all of the clusters for your account in the specified Region.

listClustersResponse_httpStatus :: Lens' ListClustersResponse Int

The response's http status code.