Network.AWS.EKS.ListUpdates

Description

Lists the updates associated with an Amazon EKS cluster or managed node group in your Amazon Web Services account, in the specified Region.

This operation returns paginated results.

Creating a Request

data ListUpdates

See: newListUpdates smart constructor.

Constructors

ListUpdates'

Fields

nextToken :: Maybe Text

The nextToken value returned from a previous paginated ListUpdates request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value.

maxResults :: Maybe Natural

The maximum number of update results returned by ListUpdates in paginated output. When you use this parameter, ListUpdates returns only maxResults results in a single page along with a nextToken response element. You can see the remaining results of the initial request by sending another ListUpdates request with the returned nextToken value. This value can be between 1 and 100. If you don't use this parameter, ListUpdates returns up to 100 results and a nextToken value if applicable.

nodegroupName :: Maybe Text

The name of the Amazon EKS managed node group to list updates for.

addonName :: Maybe Text

The names of the installed add-ons that have available updates.

name :: Text

The name of the Amazon EKS cluster to list updates for.

newListUpdates

Arguments

:: Text
$sel:name:ListUpdates'

-> ListUpdates

Create a value of ListUpdates with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListUpdates', listUpdates_nextToken - The nextToken value returned from a previous paginated ListUpdates request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value.

$sel:maxResults:ListUpdates', listUpdates_maxResults - The maximum number of update results returned by ListUpdates in paginated output. When you use this parameter, ListUpdates returns only maxResults results in a single page along with a nextToken response element. You can see the remaining results of the initial request by sending another ListUpdates request with the returned nextToken value. This value can be between 1 and 100. If you don't use this parameter, ListUpdates returns up to 100 results and a nextToken value if applicable.

$sel:nodegroupName:ListUpdates', listUpdates_nodegroupName - The name of the Amazon EKS managed node group to list updates for.

$sel:addonName:ListUpdates', listUpdates_addonName - The names of the installed add-ons that have available updates.

$sel:name:ListUpdates', listUpdates_name - The name of the Amazon EKS cluster to list updates for.

Request Lenses

listUpdates_nextToken :: Lens' ListUpdates (Maybe Text)

The nextToken value returned from a previous paginated ListUpdates request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value.

listUpdates_maxResults :: Lens' ListUpdates (Maybe Natural)

The maximum number of update results returned by ListUpdates in paginated output. When you use this parameter, ListUpdates returns only maxResults results in a single page along with a nextToken response element. You can see the remaining results of the initial request by sending another ListUpdates request with the returned nextToken value. This value can be between 1 and 100. If you don't use this parameter, ListUpdates returns up to 100 results and a nextToken value if applicable.

listUpdates_nodegroupName :: Lens' ListUpdates (Maybe Text)

The name of the Amazon EKS managed node group to list updates for.

listUpdates_addonName :: Lens' ListUpdates (Maybe Text)

The names of the installed add-ons that have available updates.

listUpdates_name :: Lens' ListUpdates Text

The name of the Amazon EKS cluster to list updates for.

Destructuring the Response

data ListUpdatesResponse

See: newListUpdatesResponse smart constructor.

Constructors

ListUpdatesResponse'

Fields

updateIds :: Maybe [Text]

A list of all the updates for the specified cluster and Region.

nextToken :: Maybe Text

The nextToken value to include in a future ListUpdates request. When the results of a ListUpdates request exceed maxResults, you can use this value to retrieve the next page of results. This value is null when there are no more results to return.

httpStatus :: Int

The response's http status code.

newListUpdatesResponse

Arguments

:: Int
$sel:httpStatus:ListUpdatesResponse'

-> ListUpdatesResponse

Create a value of ListUpdatesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:updateIds:ListUpdatesResponse', listUpdatesResponse_updateIds - A list of all the updates for the specified cluster and Region.

$sel:nextToken:ListUpdates', listUpdatesResponse_nextToken - The nextToken value to include in a future ListUpdates request. When the results of a ListUpdates request exceed maxResults, you can use this value to retrieve the next page of results. This value is null when there are no more results to return.

$sel:httpStatus:ListUpdatesResponse', listUpdatesResponse_httpStatus - The response's http status code.

Response Lenses

listUpdatesResponse_updateIds :: Lens' ListUpdatesResponse (Maybe [Text])

A list of all the updates for the specified cluster and Region.

listUpdatesResponse_nextToken :: Lens' ListUpdatesResponse (Maybe Text)

The nextToken value to include in a future ListUpdates request. When the results of a ListUpdates request exceed maxResults, you can use this value to retrieve the next page of results. This value is null when there are no more results to return.

listUpdatesResponse_httpStatus :: Lens' ListUpdatesResponse Int

The response's http status code.