Network.AWS.EKS.RegisterCluster

Description

Connects a Kubernetes cluster to the Amazon EKS control plane.

Any Kubernetes cluster can be connected to the Amazon EKS control plane to view current information about the cluster and its nodes.

Cluster connection requires two steps. First, send a RegisterClusterRequest to add it to the Amazon EKS control plane.

Second, a Manifest containing the activationID and activationCode must be applied to the Kubernetes cluster through it's native provider to provide visibility.

After the Manifest is updated and applied, then the connected cluster is visible to the Amazon EKS control plane. If the Manifest is not applied within a set amount of time, then the connected cluster will no longer be visible and must be deregistered. See DeregisterCluster.

Creating a Request

data RegisterCluster

See: newRegisterCluster smart constructor.

Constructors

RegisterCluster'

Fields

clientRequestToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

name :: Text

Define a unique name for this cluster within your AWS account.

connectorConfig :: ConnectorConfigRequest

The configuration settings required to connect the Kubernetes cluster to the Amazon EKS control plane.

newRegisterCluster

Arguments

:: Text
$sel:name:RegisterCluster'

-> ConnectorConfigRequest
$sel:connectorConfig:RegisterCluster'

-> RegisterCluster

Create a value of RegisterCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientRequestToken:RegisterCluster', registerCluster_clientRequestToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

$sel:name:RegisterCluster', registerCluster_name - Define a unique name for this cluster within your AWS account.

$sel:connectorConfig:RegisterCluster', registerCluster_connectorConfig - The configuration settings required to connect the Kubernetes cluster to the Amazon EKS control plane.

Request Lenses

registerCluster_clientRequestToken :: Lens' RegisterCluster (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

registerCluster_name :: Lens' RegisterCluster Text

Define a unique name for this cluster within your AWS account.

registerCluster_connectorConfig :: Lens' RegisterCluster ConnectorConfigRequest

The configuration settings required to connect the Kubernetes cluster to the Amazon EKS control plane.

Destructuring the Response

data RegisterClusterResponse

See: newRegisterClusterResponse smart constructor.

Constructors

RegisterClusterResponse'

Fields

cluster :: Maybe Cluster

httpStatus :: Int

The response's http status code.

newRegisterClusterResponse

Arguments

:: Int
$sel:httpStatus:RegisterClusterResponse'

-> RegisterClusterResponse

Create a value of RegisterClusterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cluster:RegisterClusterResponse', registerClusterResponse_cluster - Undocumented member.

$sel:httpStatus:RegisterClusterResponse', registerClusterResponse_httpStatus - The response's http status code.

Response Lenses

registerClusterResponse_cluster :: Lens' RegisterClusterResponse (Maybe Cluster)

Undocumented member.

registerClusterResponse_httpStatus :: Lens' RegisterClusterResponse Int

The response's http status code.