Network.AWS.EKS.Types.Addon

Description

Documentation

data Addon

An Amazon EKS add-on.

See: newAddon smart constructor.

Constructors

Addon'

Fields

status :: Maybe AddonStatus

The status of the add-on.

modifiedAt :: Maybe POSIX

The date and time that the add-on was last modified.

serviceAccountRoleArn :: Maybe Text

The Amazon Resource Name (ARN) of the IAM role that is bound to the Kubernetes service account used by the add-on.

createdAt :: Maybe POSIX

The date and time that the add-on was created.

addonArn :: Maybe Text

The Amazon Resource Name (ARN) of the add-on.

addonVersion :: Maybe Text

The version of the add-on.

addonName :: Maybe Text

The name of the add-on.

health :: Maybe AddonHealth

An object that represents the health of the add-on.

tags :: Maybe (HashMap Text Text)

The metadata that you apply to the add-on to assist with categorization and organization. Each tag consists of a key and an optional value, both of which you define. Add-on tags do not propagate to any other resources associated with the cluster.

clusterName :: Maybe Text

The name of the cluster.

newAddon :: Addon

Create a value of Addon with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Addon', addon_status - The status of the add-on.

$sel:modifiedAt:Addon', addon_modifiedAt - The date and time that the add-on was last modified.

$sel:serviceAccountRoleArn:Addon', addon_serviceAccountRoleArn - The Amazon Resource Name (ARN) of the IAM role that is bound to the Kubernetes service account used by the add-on.

$sel:createdAt:Addon', addon_createdAt - The date and time that the add-on was created.

$sel:addonArn:Addon', addon_addonArn - The Amazon Resource Name (ARN) of the add-on.

$sel:addonVersion:Addon', addon_addonVersion - The version of the add-on.

$sel:addonName:Addon', addon_addonName - The name of the add-on.

$sel:health:Addon', addon_health - An object that represents the health of the add-on.

$sel:tags:Addon', addon_tags - The metadata that you apply to the add-on to assist with categorization and organization. Each tag consists of a key and an optional value, both of which you define. Add-on tags do not propagate to any other resources associated with the cluster.

$sel:clusterName:Addon', addon_clusterName - The name of the cluster.

addon_status :: Lens' Addon (Maybe AddonStatus)

The status of the add-on.

addon_modifiedAt :: Lens' Addon (Maybe UTCTime)

The date and time that the add-on was last modified.

addon_serviceAccountRoleArn :: Lens' Addon (Maybe Text)

The Amazon Resource Name (ARN) of the IAM role that is bound to the Kubernetes service account used by the add-on.

addon_createdAt :: Lens' Addon (Maybe UTCTime)

The date and time that the add-on was created.

addon_addonArn :: Lens' Addon (Maybe Text)

The Amazon Resource Name (ARN) of the add-on.

addon_addonVersion :: Lens' Addon (Maybe Text)

The version of the add-on.

addon_addonName :: Lens' Addon (Maybe Text)

The name of the add-on.

addon_health :: Lens' Addon (Maybe AddonHealth)

An object that represents the health of the add-on.

addon_tags :: Lens' Addon (Maybe (HashMap Text Text))

The metadata that you apply to the add-on to assist with categorization and organization. Each tag consists of a key and an optional value, both of which you define. Add-on tags do not propagate to any other resources associated with the cluster.

addon_clusterName :: Lens' Addon (Maybe Text)

The name of the cluster.