Network.AWS.EKS.Types.AddonVersionInfo

Description

Documentation

data AddonVersionInfo

Information about an add-on version.

See: newAddonVersionInfo smart constructor.

Constructors

AddonVersionInfo'

Fields

addonVersion :: Maybe Text

The version of the add-on.

architecture :: Maybe [Text]

The architectures that the version supports.

compatibilities :: Maybe [Compatibility]

An object that represents the compatibilities of a version.

newAddonVersionInfo :: AddonVersionInfo

Create a value of AddonVersionInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:addonVersion:AddonVersionInfo', addonVersionInfo_addonVersion - The version of the add-on.

$sel:architecture:AddonVersionInfo', addonVersionInfo_architecture - The architectures that the version supports.

$sel:compatibilities:AddonVersionInfo', addonVersionInfo_compatibilities - An object that represents the compatibilities of a version.

addonVersionInfo_addonVersion :: Lens' AddonVersionInfo (Maybe Text)

The version of the add-on.

addonVersionInfo_architecture :: Lens' AddonVersionInfo (Maybe [Text])

The architectures that the version supports.

addonVersionInfo_compatibilities :: Lens' AddonVersionInfo (Maybe [Compatibility])

An object that represents the compatibilities of a version.