Network.AWS.EKS.Types.Cluster

Description

Documentation

data Cluster

An object representing an Amazon EKS cluster.

See: newCluster smart constructor.

Constructors

Cluster'

Fields

status :: Maybe ClusterStatus

The current status of the cluster.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the cluster.

createdAt :: Maybe POSIX

The Unix epoch timestamp in seconds for when the cluster was created.

platformVersion :: Maybe Text

The platform version of your Amazon EKS cluster. For more information, see Platform Versions in the /Amazon EKS User Guide/ .

kubernetesNetworkConfig :: Maybe KubernetesNetworkConfigResponse

The Kubernetes network configuration for the cluster.

connectorConfig :: Maybe ConnectorConfigResponse

The configuration used to connect to a cluster for registration.

certificateAuthority :: Maybe Certificate

The certificate-authority-data for your cluster.

name :: Maybe Text

The name of the cluster.

version :: Maybe Text

The Kubernetes server version for the cluster.

encryptionConfig :: Maybe [EncryptionConfig]

The encryption configuration for the cluster.

endpoint :: Maybe Text

The endpoint for your Kubernetes API server.

clientRequestToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

logging :: Maybe Logging

The logging configuration for your cluster.

identity :: Maybe Identity

The identity provider information for the cluster.

resourcesVpcConfig :: Maybe VpcConfigResponse

The VPC configuration used by the cluster control plane. Amazon EKS VPC resources have specific requirements to work properly with Kubernetes. For more information, see Cluster VPC Considerations and Cluster Security Group Considerations in the Amazon EKS User Guide.

tags :: Maybe (HashMap Text Text)

The metadata that you apply to the cluster to assist with categorization and organization. Each tag consists of a key and an optional value, both of which you define. Cluster tags do not propagate to any other resources associated with the cluster.

roleArn :: Maybe Text

The Amazon Resource Name (ARN) of the IAM role that provides permissions for the Kubernetes control plane to make calls to Amazon Web Services API operations on your behalf.

newCluster :: Cluster

Create a value of Cluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Cluster', cluster_status - The current status of the cluster.

$sel:arn:Cluster', cluster_arn - The Amazon Resource Name (ARN) of the cluster.

$sel:createdAt:Cluster', cluster_createdAt - The Unix epoch timestamp in seconds for when the cluster was created.

$sel:platformVersion:Cluster', cluster_platformVersion - The platform version of your Amazon EKS cluster. For more information, see Platform Versions in the /Amazon EKS User Guide/ .

$sel:kubernetesNetworkConfig:Cluster', cluster_kubernetesNetworkConfig - The Kubernetes network configuration for the cluster.

$sel:connectorConfig:Cluster', cluster_connectorConfig - The configuration used to connect to a cluster for registration.

$sel:certificateAuthority:Cluster', cluster_certificateAuthority - The certificate-authority-data for your cluster.

$sel:name:Cluster', cluster_name - The name of the cluster.

$sel:version:Cluster', cluster_version - The Kubernetes server version for the cluster.

$sel:encryptionConfig:Cluster', cluster_encryptionConfig - The encryption configuration for the cluster.

$sel:endpoint:Cluster', cluster_endpoint - The endpoint for your Kubernetes API server.

$sel:clientRequestToken:Cluster', cluster_clientRequestToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

$sel:logging:Cluster', cluster_logging - The logging configuration for your cluster.

$sel:identity:Cluster', cluster_identity - The identity provider information for the cluster.

$sel:resourcesVpcConfig:Cluster', cluster_resourcesVpcConfig - The VPC configuration used by the cluster control plane. Amazon EKS VPC resources have specific requirements to work properly with Kubernetes. For more information, see Cluster VPC Considerations and Cluster Security Group Considerations in the Amazon EKS User Guide.

$sel:tags:Cluster', cluster_tags - The metadata that you apply to the cluster to assist with categorization and organization. Each tag consists of a key and an optional value, both of which you define. Cluster tags do not propagate to any other resources associated with the cluster.

$sel:roleArn:Cluster', cluster_roleArn - The Amazon Resource Name (ARN) of the IAM role that provides permissions for the Kubernetes control plane to make calls to Amazon Web Services API operations on your behalf.

cluster_status :: Lens' Cluster (Maybe ClusterStatus)

The current status of the cluster.

cluster_arn :: Lens' Cluster (Maybe Text)

The Amazon Resource Name (ARN) of the cluster.

cluster_createdAt :: Lens' Cluster (Maybe UTCTime)

The Unix epoch timestamp in seconds for when the cluster was created.

cluster_platformVersion :: Lens' Cluster (Maybe Text)

The platform version of your Amazon EKS cluster. For more information, see Platform Versions in the /Amazon EKS User Guide/ .

cluster_kubernetesNetworkConfig :: Lens' Cluster (Maybe KubernetesNetworkConfigResponse)

The Kubernetes network configuration for the cluster.

cluster_connectorConfig :: Lens' Cluster (Maybe ConnectorConfigResponse)

The configuration used to connect to a cluster for registration.

cluster_certificateAuthority :: Lens' Cluster (Maybe Certificate)

The certificate-authority-data for your cluster.

cluster_name :: Lens' Cluster (Maybe Text)

The name of the cluster.

cluster_version :: Lens' Cluster (Maybe Text)

The Kubernetes server version for the cluster.

cluster_encryptionConfig :: Lens' Cluster (Maybe [EncryptionConfig])

The encryption configuration for the cluster.

cluster_endpoint :: Lens' Cluster (Maybe Text)

The endpoint for your Kubernetes API server.

cluster_clientRequestToken :: Lens' Cluster (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

cluster_logging :: Lens' Cluster (Maybe Logging)

The logging configuration for your cluster.

cluster_identity :: Lens' Cluster (Maybe Identity)

The identity provider information for the cluster.

cluster_resourcesVpcConfig :: Lens' Cluster (Maybe VpcConfigResponse)

The VPC configuration used by the cluster control plane. Amazon EKS VPC resources have specific requirements to work properly with Kubernetes. For more information, see Cluster VPC Considerations and Cluster Security Group Considerations in the Amazon EKS User Guide.

cluster_tags :: Lens' Cluster (Maybe (HashMap Text Text))

The metadata that you apply to the cluster to assist with categorization and organization. Each tag consists of a key and an optional value, both of which you define. Cluster tags do not propagate to any other resources associated with the cluster.

cluster_roleArn :: Lens' Cluster (Maybe Text)

The Amazon Resource Name (ARN) of the IAM role that provides permissions for the Kubernetes control plane to make calls to Amazon Web Services API operations on your behalf.