Network.AWS.EKS.Types.ConnectorConfigResponse

Description

Documentation

data ConnectorConfigResponse

The full description of your connected cluster.

See: newConnectorConfigResponse smart constructor.

Constructors

ConnectorConfigResponse'

Fields

activationCode :: Maybe Text

A unique code associated with the cluster for registration purposes.

roleArn :: Maybe Text

The Amazon Resource Name (ARN) of the role that is used by the EKS connector to communicate with AWS services from the connected Kubernetes cluster.

activationId :: Maybe Text

A unique ID associated with the cluster for registration purposes.

provider :: Maybe Text

The cluster's cloud service provider.

activationExpiry :: Maybe POSIX

The expiration time of the connected cluster. The cluster's YAML file must be applied through the native provider.

newConnectorConfigResponse :: ConnectorConfigResponse

Create a value of ConnectorConfigResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:activationCode:ConnectorConfigResponse', connectorConfigResponse_activationCode - A unique code associated with the cluster for registration purposes.

$sel:roleArn:ConnectorConfigResponse', connectorConfigResponse_roleArn - The Amazon Resource Name (ARN) of the role that is used by the EKS connector to communicate with AWS services from the connected Kubernetes cluster.

$sel:activationId:ConnectorConfigResponse', connectorConfigResponse_activationId - A unique ID associated with the cluster for registration purposes.

$sel:provider:ConnectorConfigResponse', connectorConfigResponse_provider - The cluster's cloud service provider.

$sel:activationExpiry:ConnectorConfigResponse', connectorConfigResponse_activationExpiry - The expiration time of the connected cluster. The cluster's YAML file must be applied through the native provider.

connectorConfigResponse_activationCode :: Lens' ConnectorConfigResponse (Maybe Text)

A unique code associated with the cluster for registration purposes.

connectorConfigResponse_roleArn :: Lens' ConnectorConfigResponse (Maybe Text)

The Amazon Resource Name (ARN) of the role that is used by the EKS connector to communicate with AWS services from the connected Kubernetes cluster.

connectorConfigResponse_activationId :: Lens' ConnectorConfigResponse (Maybe Text)

A unique ID associated with the cluster for registration purposes.

connectorConfigResponse_provider :: Lens' ConnectorConfigResponse (Maybe Text)

The cluster's cloud service provider.

connectorConfigResponse_activationExpiry :: Lens' ConnectorConfigResponse (Maybe UTCTime)

The expiration time of the connected cluster. The cluster's YAML file must be applied through the native provider.