Network.AWS.EKS.Types.FargateProfile

Description

Documentation

data FargateProfile

An object representing an Fargate profile.

See: newFargateProfile smart constructor.

Constructors

FargateProfile'

Fields

fargateProfileArn :: Maybe Text

The full Amazon Resource Name (ARN) of the Fargate profile.

status :: Maybe FargateProfileStatus

The current status of the Fargate profile.

createdAt :: Maybe POSIX

The Unix epoch timestamp in seconds for when the Fargate profile was created.

subnets :: Maybe [Text]

The IDs of subnets to launch pods into.

clusterName :: Maybe Text

The name of the Amazon EKS cluster that the Fargate profile belongs to.

podExecutionRoleArn :: Maybe Text

The Amazon Resource Name (ARN) of the pod execution role to use for pods that match the selectors in the Fargate profile. For more information, see Pod Execution Role in the Amazon EKS User Guide.

fargateProfileName :: Maybe Text

The name of the Fargate profile.

selectors :: Maybe [FargateProfileSelector]

The selectors to match for pods to use this Fargate profile.

tags :: Maybe (HashMap Text Text)

The metadata applied to the Fargate profile to assist with categorization and organization. Each tag consists of a key and an optional value, both of which you define. Fargate profile tags do not propagate to any other resources associated with the Fargate profile, such as the pods that are scheduled with it.

newFargateProfile :: FargateProfile

Create a value of FargateProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fargateProfileArn:FargateProfile', fargateProfile_fargateProfileArn - The full Amazon Resource Name (ARN) of the Fargate profile.

$sel:status:FargateProfile', fargateProfile_status - The current status of the Fargate profile.

$sel:createdAt:FargateProfile', fargateProfile_createdAt - The Unix epoch timestamp in seconds for when the Fargate profile was created.

$sel:subnets:FargateProfile', fargateProfile_subnets - The IDs of subnets to launch pods into.

$sel:clusterName:FargateProfile', fargateProfile_clusterName - The name of the Amazon EKS cluster that the Fargate profile belongs to.

$sel:podExecutionRoleArn:FargateProfile', fargateProfile_podExecutionRoleArn - The Amazon Resource Name (ARN) of the pod execution role to use for pods that match the selectors in the Fargate profile. For more information, see Pod Execution Role in the Amazon EKS User Guide.

$sel:fargateProfileName:FargateProfile', fargateProfile_fargateProfileName - The name of the Fargate profile.

$sel:selectors:FargateProfile', fargateProfile_selectors - The selectors to match for pods to use this Fargate profile.

$sel:tags:FargateProfile', fargateProfile_tags - The metadata applied to the Fargate profile to assist with categorization and organization. Each tag consists of a key and an optional value, both of which you define. Fargate profile tags do not propagate to any other resources associated with the Fargate profile, such as the pods that are scheduled with it.

fargateProfile_fargateProfileArn :: Lens' FargateProfile (Maybe Text)

The full Amazon Resource Name (ARN) of the Fargate profile.

fargateProfile_status :: Lens' FargateProfile (Maybe FargateProfileStatus)

The current status of the Fargate profile.

fargateProfile_createdAt :: Lens' FargateProfile (Maybe UTCTime)

The Unix epoch timestamp in seconds for when the Fargate profile was created.

fargateProfile_subnets :: Lens' FargateProfile (Maybe [Text])

The IDs of subnets to launch pods into.

fargateProfile_clusterName :: Lens' FargateProfile (Maybe Text)

The name of the Amazon EKS cluster that the Fargate profile belongs to.

fargateProfile_podExecutionRoleArn :: Lens' FargateProfile (Maybe Text)

The Amazon Resource Name (ARN) of the pod execution role to use for pods that match the selectors in the Fargate profile. For more information, see Pod Execution Role in the Amazon EKS User Guide.

fargateProfile_fargateProfileName :: Lens' FargateProfile (Maybe Text)

The name of the Fargate profile.

fargateProfile_selectors :: Lens' FargateProfile (Maybe [FargateProfileSelector])

The selectors to match for pods to use this Fargate profile.

fargateProfile_tags :: Lens' FargateProfile (Maybe (HashMap Text Text))

The metadata applied to the Fargate profile to assist with categorization and organization. Each tag consists of a key and an optional value, both of which you define. Fargate profile tags do not propagate to any other resources associated with the Fargate profile, such as the pods that are scheduled with it.