Network.AWS.EKS.Types.Nodegroup

Description

Documentation

data Nodegroup

An object representing an Amazon EKS managed node group.

See: newNodegroup smart constructor.

Constructors

Nodegroup'

Fields

scalingConfig :: Maybe NodegroupScalingConfig

The scaling configuration details for the Auto Scaling group that is associated with your node group.

capacityType :: Maybe CapacityTypes

The capacity type of your managed node group.

status :: Maybe NodegroupStatus

The current status of the managed node group.

modifiedAt :: Maybe POSIX

The Unix epoch timestamp in seconds for when the managed node group was last modified.

releaseVersion :: Maybe Text

If the node group was deployed using a launch template with a custom AMI, then this is the AMI ID that was specified in the launch template. For node groups that weren't deployed using a launch template, this is the version of the Amazon EKS optimized AMI that the node group was deployed with.

nodegroupName :: Maybe Text

The name associated with an Amazon EKS managed node group.

remoteAccess :: Maybe RemoteAccessConfig

If the node group wasn't deployed with a launch template, then this is the remote access configuration that is associated with the node group. If the node group was deployed with a launch template, then this is null.

diskSize :: Maybe Int

If the node group wasn't deployed with a launch template, then this is the disk size in the node group configuration. If the node group was deployed with a launch template, then this is null.

createdAt :: Maybe POSIX

The Unix epoch timestamp in seconds for when the managed node group was created.

labels :: Maybe (HashMap Text Text)

The Kubernetes labels applied to the nodes in the node group.

Only labels that are applied with the Amazon EKS API are shown here. There may be other Kubernetes labels applied to the nodes in this group.

launchTemplate :: Maybe LaunchTemplateSpecification

If a launch template was used to create the node group, then this is the launch template that was used.

version :: Maybe Text

The Kubernetes version of the managed node group.

nodeRole :: Maybe Text

The IAM role associated with your node group. The Amazon EKS node kubelet daemon makes calls to Amazon Web Services APIs on your behalf. Nodes receive permissions for these API calls through an IAM instance profile and associated policies.

health :: Maybe NodegroupHealth

The health status of the node group. If there are issues with your node group's health, they are listed here.

resources :: Maybe NodegroupResources

The resources associated with the node group, such as Auto Scaling groups and security groups for remote access.

tags :: Maybe (HashMap Text Text)

The metadata applied to the node group to assist with categorization and organization. Each tag consists of a key and an optional value, both of which you define. Node group tags do not propagate to any other resources associated with the node group, such as the Amazon EC2 instances or subnets.

updateConfig :: Maybe NodegroupUpdateConfig

The node group update configuration.

subnets :: Maybe [Text]

The subnets that were specified for the Auto Scaling group that is associated with your node group.

clusterName :: Maybe Text

The name of the cluster that the managed node group resides in.

amiType :: Maybe AMITypes

If the node group was deployed using a launch template with a custom AMI, then this is CUSTOM. For node groups that weren't deployed using a launch template, this is the AMI type that was specified in the node group configuration.

taints :: Maybe [Taint]

The Kubernetes taints to be applied to the nodes in the node group when they are created. Effect is one of No_Schedule, Prefer_No_Schedule, or No_Execute. Kubernetes taints can be used together with tolerations to control how workloads are scheduled to your nodes.

instanceTypes :: Maybe [Text]

If the node group wasn't deployed with a launch template, then this is the instance type that is associated with the node group. If the node group was deployed with a launch template, then this is null.

nodegroupArn :: Maybe Text

The Amazon Resource Name (ARN) associated with the managed node group.

newNodegroup :: Nodegroup

Create a value of Nodegroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:scalingConfig:Nodegroup', nodegroup_scalingConfig - The scaling configuration details for the Auto Scaling group that is associated with your node group.

$sel:capacityType:Nodegroup', nodegroup_capacityType - The capacity type of your managed node group.

$sel:status:Nodegroup', nodegroup_status - The current status of the managed node group.

$sel:modifiedAt:Nodegroup', nodegroup_modifiedAt - The Unix epoch timestamp in seconds for when the managed node group was last modified.

$sel:releaseVersion:Nodegroup', nodegroup_releaseVersion - If the node group was deployed using a launch template with a custom AMI, then this is the AMI ID that was specified in the launch template. For node groups that weren't deployed using a launch template, this is the version of the Amazon EKS optimized AMI that the node group was deployed with.

$sel:nodegroupName:Nodegroup', nodegroup_nodegroupName - The name associated with an Amazon EKS managed node group.

$sel:remoteAccess:Nodegroup', nodegroup_remoteAccess - If the node group wasn't deployed with a launch template, then this is the remote access configuration that is associated with the node group. If the node group was deployed with a launch template, then this is null.

$sel:diskSize:Nodegroup', nodegroup_diskSize - If the node group wasn't deployed with a launch template, then this is the disk size in the node group configuration. If the node group was deployed with a launch template, then this is null.

$sel:createdAt:Nodegroup', nodegroup_createdAt - The Unix epoch timestamp in seconds for when the managed node group was created.

$sel:labels:Nodegroup', nodegroup_labels - The Kubernetes labels applied to the nodes in the node group.

Only labels that are applied with the Amazon EKS API are shown here. There may be other Kubernetes labels applied to the nodes in this group.

$sel:launchTemplate:Nodegroup', nodegroup_launchTemplate - If a launch template was used to create the node group, then this is the launch template that was used.

$sel:version:Nodegroup', nodegroup_version - The Kubernetes version of the managed node group.

$sel:nodeRole:Nodegroup', nodegroup_nodeRole - The IAM role associated with your node group. The Amazon EKS node kubelet daemon makes calls to Amazon Web Services APIs on your behalf. Nodes receive permissions for these API calls through an IAM instance profile and associated policies.

$sel:health:Nodegroup', nodegroup_health - The health status of the node group. If there are issues with your node group's health, they are listed here.

$sel:resources:Nodegroup', nodegroup_resources - The resources associated with the node group, such as Auto Scaling groups and security groups for remote access.

$sel:tags:Nodegroup', nodegroup_tags - The metadata applied to the node group to assist with categorization and organization. Each tag consists of a key and an optional value, both of which you define. Node group tags do not propagate to any other resources associated with the node group, such as the Amazon EC2 instances or subnets.

$sel:updateConfig:Nodegroup', nodegroup_updateConfig - The node group update configuration.

$sel:subnets:Nodegroup', nodegroup_subnets - The subnets that were specified for the Auto Scaling group that is associated with your node group.

$sel:clusterName:Nodegroup', nodegroup_clusterName - The name of the cluster that the managed node group resides in.

$sel:amiType:Nodegroup', nodegroup_amiType - If the node group was deployed using a launch template with a custom AMI, then this is CUSTOM. For node groups that weren't deployed using a launch template, this is the AMI type that was specified in the node group configuration.

$sel:taints:Nodegroup', nodegroup_taints - The Kubernetes taints to be applied to the nodes in the node group when they are created. Effect is one of No_Schedule, Prefer_No_Schedule, or No_Execute. Kubernetes taints can be used together with tolerations to control how workloads are scheduled to your nodes.

$sel:instanceTypes:Nodegroup', nodegroup_instanceTypes - If the node group wasn't deployed with a launch template, then this is the instance type that is associated with the node group. If the node group was deployed with a launch template, then this is null.

$sel:nodegroupArn:Nodegroup', nodegroup_nodegroupArn - The Amazon Resource Name (ARN) associated with the managed node group.

nodegroup_scalingConfig :: Lens' Nodegroup (Maybe NodegroupScalingConfig)

The scaling configuration details for the Auto Scaling group that is associated with your node group.

nodegroup_capacityType :: Lens' Nodegroup (Maybe CapacityTypes)

The capacity type of your managed node group.

nodegroup_status :: Lens' Nodegroup (Maybe NodegroupStatus)

The current status of the managed node group.

nodegroup_modifiedAt :: Lens' Nodegroup (Maybe UTCTime)

The Unix epoch timestamp in seconds for when the managed node group was last modified.

nodegroup_releaseVersion :: Lens' Nodegroup (Maybe Text)

If the node group was deployed using a launch template with a custom AMI, then this is the AMI ID that was specified in the launch template. For node groups that weren't deployed using a launch template, this is the version of the Amazon EKS optimized AMI that the node group was deployed with.

nodegroup_nodegroupName :: Lens' Nodegroup (Maybe Text)

The name associated with an Amazon EKS managed node group.

nodegroup_remoteAccess :: Lens' Nodegroup (Maybe RemoteAccessConfig)

If the node group wasn't deployed with a launch template, then this is the remote access configuration that is associated with the node group. If the node group was deployed with a launch template, then this is null.

nodegroup_diskSize :: Lens' Nodegroup (Maybe Int)

If the node group wasn't deployed with a launch template, then this is the disk size in the node group configuration. If the node group was deployed with a launch template, then this is null.

nodegroup_createdAt :: Lens' Nodegroup (Maybe UTCTime)

The Unix epoch timestamp in seconds for when the managed node group was created.

nodegroup_labels :: Lens' Nodegroup (Maybe (HashMap Text Text))

The Kubernetes labels applied to the nodes in the node group.

Only labels that are applied with the Amazon EKS API are shown here. There may be other Kubernetes labels applied to the nodes in this group.

nodegroup_launchTemplate :: Lens' Nodegroup (Maybe LaunchTemplateSpecification)

If a launch template was used to create the node group, then this is the launch template that was used.

nodegroup_version :: Lens' Nodegroup (Maybe Text)

The Kubernetes version of the managed node group.

nodegroup_nodeRole :: Lens' Nodegroup (Maybe Text)

The IAM role associated with your node group. The Amazon EKS node kubelet daemon makes calls to Amazon Web Services APIs on your behalf. Nodes receive permissions for these API calls through an IAM instance profile and associated policies.

nodegroup_health :: Lens' Nodegroup (Maybe NodegroupHealth)

The health status of the node group. If there are issues with your node group's health, they are listed here.

nodegroup_resources :: Lens' Nodegroup (Maybe NodegroupResources)

The resources associated with the node group, such as Auto Scaling groups and security groups for remote access.

nodegroup_tags :: Lens' Nodegroup (Maybe (HashMap Text Text))

The metadata applied to the node group to assist with categorization and organization. Each tag consists of a key and an optional value, both of which you define. Node group tags do not propagate to any other resources associated with the node group, such as the Amazon EC2 instances or subnets.

nodegroup_updateConfig :: Lens' Nodegroup (Maybe NodegroupUpdateConfig)

The node group update configuration.

nodegroup_subnets :: Lens' Nodegroup (Maybe [Text])

The subnets that were specified for the Auto Scaling group that is associated with your node group.

nodegroup_clusterName :: Lens' Nodegroup (Maybe Text)

The name of the cluster that the managed node group resides in.

nodegroup_amiType :: Lens' Nodegroup (Maybe AMITypes)

If the node group was deployed using a launch template with a custom AMI, then this is CUSTOM. For node groups that weren't deployed using a launch template, this is the AMI type that was specified in the node group configuration.

nodegroup_taints :: Lens' Nodegroup (Maybe [Taint])

The Kubernetes taints to be applied to the nodes in the node group when they are created. Effect is one of No_Schedule, Prefer_No_Schedule, or No_Execute. Kubernetes taints can be used together with tolerations to control how workloads are scheduled to your nodes.

nodegroup_instanceTypes :: Lens' Nodegroup (Maybe [Text])

If the node group wasn't deployed with a launch template, then this is the instance type that is associated with the node group. If the node group was deployed with a launch template, then this is null.

nodegroup_nodegroupArn :: Lens' Nodegroup (Maybe Text)

The Amazon Resource Name (ARN) associated with the managed node group.