Network.AWS.EKS.Types.NodegroupHealth

Description

Documentation

data NodegroupHealth

An object representing the health status of the node group.

See: newNodegroupHealth smart constructor.

Constructors

NodegroupHealth'

Fields

issues :: Maybe [Issue]

Any issues that are associated with the node group.

newNodegroupHealth :: NodegroupHealth

Create a value of NodegroupHealth with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:issues:NodegroupHealth', nodegroupHealth_issues - Any issues that are associated with the node group.

nodegroupHealth_issues :: Lens' NodegroupHealth (Maybe [Issue])

Any issues that are associated with the node group.