Network.AWS.EKS.Types.NodegroupScalingConfig

Description

Documentation

data NodegroupScalingConfig

An object representing the scaling configuration details for the Auto Scaling group that is associated with your node group. When creating a node group, you must specify all or none of the properties. When updating a node group, you can specify any or none of the properties.

See: newNodegroupScalingConfig smart constructor.

Constructors

NodegroupScalingConfig'

Fields

desiredSize :: Maybe Natural

The current number of nodes that the managed node group should maintain.

maxSize :: Maybe Natural

The maximum number of nodes that the managed node group can scale out to. For information about the maximum number that you can specify, see Amazon EKS service quotas in the Amazon EKS User Guide.

minSize :: Maybe Natural

The minimum number of nodes that the managed node group can scale in to.

newNodegroupScalingConfig :: NodegroupScalingConfig

Create a value of NodegroupScalingConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:desiredSize:NodegroupScalingConfig', nodegroupScalingConfig_desiredSize - The current number of nodes that the managed node group should maintain.

$sel:maxSize:NodegroupScalingConfig', nodegroupScalingConfig_maxSize - The maximum number of nodes that the managed node group can scale out to. For information about the maximum number that you can specify, see Amazon EKS service quotas in the Amazon EKS User Guide.

$sel:minSize:NodegroupScalingConfig', nodegroupScalingConfig_minSize - The minimum number of nodes that the managed node group can scale in to.

nodegroupScalingConfig_desiredSize :: Lens' NodegroupScalingConfig (Maybe Natural)

The current number of nodes that the managed node group should maintain.

nodegroupScalingConfig_maxSize :: Lens' NodegroupScalingConfig (Maybe Natural)

The maximum number of nodes that the managed node group can scale out to. For information about the maximum number that you can specify, see Amazon EKS service quotas in the Amazon EKS User Guide.

nodegroupScalingConfig_minSize :: Lens' NodegroupScalingConfig (Maybe Natural)

The minimum number of nodes that the managed node group can scale in to.