Network.AWS.EKS.Types.VpcConfigRequest

Description

Documentation

data VpcConfigRequest

An object representing the VPC configuration to use for an Amazon EKS cluster.

See: newVpcConfigRequest smart constructor.

Constructors

VpcConfigRequest'

Fields

securityGroupIds :: Maybe [Text]

Specify one or more security groups for the cross-account elastic network interfaces that Amazon EKS creates to use that allow communication between your nodes and the Kubernetes control plane. If you don't specify any security groups, then familiarize yourself with the difference between Amazon EKS defaults for clusters deployed with Kubernetes:

1.14 Amazon EKS platform version eks.2 and earlier

1.14 Amazon EKS platform version eks.3 and later

For more information, see Amazon EKS security group considerations in the /Amazon EKS User Guide/ .

endpointPublicAccess :: Maybe Bool

Set this value to false to disable public access to your cluster's Kubernetes API server endpoint. If you disable public access, your cluster's Kubernetes API server can only receive requests from within the cluster VPC. The default value for this parameter is true, which enables public access for your Kubernetes API server. For more information, see Amazon EKS cluster endpoint access control in the /Amazon EKS User Guide/ .

subnetIds :: Maybe [Text]

Specify subnets for your Amazon EKS nodes. Amazon EKS creates cross-account elastic network interfaces in these subnets to allow communication between your nodes and the Kubernetes control plane.

endpointPrivateAccess :: Maybe Bool

Set this value to true to enable private access for your cluster's Kubernetes API server endpoint. If you enable private access, Kubernetes API requests from within your cluster's VPC use the private VPC endpoint. The default value for this parameter is false, which disables private access for your Kubernetes API server. If you disable private access and you have nodes or Fargate pods in the cluster, then ensure that publicAccessCidrs includes the necessary CIDR blocks for communication with the nodes or Fargate pods. For more information, see Amazon EKS cluster endpoint access control in the /Amazon EKS User Guide/ .

publicAccessCidrs :: Maybe [Text]

The CIDR blocks that are allowed access to your cluster's public Kubernetes API server endpoint. Communication to the endpoint from addresses outside of the CIDR blocks that you specify is denied. The default value is 0.0.0.0/0. If you've disabled private endpoint access and you have nodes or Fargate pods in the cluster, then ensure that you specify the necessary CIDR blocks. For more information, see Amazon EKS cluster endpoint access control in the /Amazon EKS User Guide/ .

newVpcConfigRequest :: VpcConfigRequest

Create a value of VpcConfigRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:securityGroupIds:VpcConfigRequest', vpcConfigRequest_securityGroupIds - Specify one or more security groups for the cross-account elastic network interfaces that Amazon EKS creates to use that allow communication between your nodes and the Kubernetes control plane. If you don't specify any security groups, then familiarize yourself with the difference between Amazon EKS defaults for clusters deployed with Kubernetes:

1.14 Amazon EKS platform version eks.2 and earlier

1.14 Amazon EKS platform version eks.3 and later

For more information, see Amazon EKS security group considerations in the /Amazon EKS User Guide/ .

$sel:endpointPublicAccess:VpcConfigRequest', vpcConfigRequest_endpointPublicAccess - Set this value to false to disable public access to your cluster's Kubernetes API server endpoint. If you disable public access, your cluster's Kubernetes API server can only receive requests from within the cluster VPC. The default value for this parameter is true, which enables public access for your Kubernetes API server. For more information, see Amazon EKS cluster endpoint access control in the /Amazon EKS User Guide/ .

$sel:subnetIds:VpcConfigRequest', vpcConfigRequest_subnetIds - Specify subnets for your Amazon EKS nodes. Amazon EKS creates cross-account elastic network interfaces in these subnets to allow communication between your nodes and the Kubernetes control plane.

$sel:endpointPrivateAccess:VpcConfigRequest', vpcConfigRequest_endpointPrivateAccess - Set this value to true to enable private access for your cluster's Kubernetes API server endpoint. If you enable private access, Kubernetes API requests from within your cluster's VPC use the private VPC endpoint. The default value for this parameter is false, which disables private access for your Kubernetes API server. If you disable private access and you have nodes or Fargate pods in the cluster, then ensure that publicAccessCidrs includes the necessary CIDR blocks for communication with the nodes or Fargate pods. For more information, see Amazon EKS cluster endpoint access control in the /Amazon EKS User Guide/ .

$sel:publicAccessCidrs:VpcConfigRequest', vpcConfigRequest_publicAccessCidrs - The CIDR blocks that are allowed access to your cluster's public Kubernetes API server endpoint. Communication to the endpoint from addresses outside of the CIDR blocks that you specify is denied. The default value is 0.0.0.0/0. If you've disabled private endpoint access and you have nodes or Fargate pods in the cluster, then ensure that you specify the necessary CIDR blocks. For more information, see Amazon EKS cluster endpoint access control in the /Amazon EKS User Guide/ .

vpcConfigRequest_securityGroupIds :: Lens' VpcConfigRequest (Maybe [Text])

Specify one or more security groups for the cross-account elastic network interfaces that Amazon EKS creates to use that allow communication between your nodes and the Kubernetes control plane. If you don't specify any security groups, then familiarize yourself with the difference between Amazon EKS defaults for clusters deployed with Kubernetes:

1.14 Amazon EKS platform version eks.2 and earlier

1.14 Amazon EKS platform version eks.3 and later

For more information, see Amazon EKS security group considerations in the /Amazon EKS User Guide/ .

vpcConfigRequest_endpointPublicAccess :: Lens' VpcConfigRequest (Maybe Bool)

Set this value to false to disable public access to your cluster's Kubernetes API server endpoint. If you disable public access, your cluster's Kubernetes API server can only receive requests from within the cluster VPC. The default value for this parameter is true, which enables public access for your Kubernetes API server. For more information, see Amazon EKS cluster endpoint access control in the /Amazon EKS User Guide/ .

vpcConfigRequest_subnetIds :: Lens' VpcConfigRequest (Maybe [Text])

Specify subnets for your Amazon EKS nodes. Amazon EKS creates cross-account elastic network interfaces in these subnets to allow communication between your nodes and the Kubernetes control plane.

vpcConfigRequest_endpointPrivateAccess :: Lens' VpcConfigRequest (Maybe Bool)

Set this value to true to enable private access for your cluster's Kubernetes API server endpoint. If you enable private access, Kubernetes API requests from within your cluster's VPC use the private VPC endpoint. The default value for this parameter is false, which disables private access for your Kubernetes API server. If you disable private access and you have nodes or Fargate pods in the cluster, then ensure that publicAccessCidrs includes the necessary CIDR blocks for communication with the nodes or Fargate pods. For more information, see Amazon EKS cluster endpoint access control in the /Amazon EKS User Guide/ .

vpcConfigRequest_publicAccessCidrs :: Lens' VpcConfigRequest (Maybe [Text])

The CIDR blocks that are allowed access to your cluster's public Kubernetes API server endpoint. Communication to the endpoint from addresses outside of the CIDR blocks that you specify is denied. The default value is 0.0.0.0/0. If you've disabled private endpoint access and you have nodes or Fargate pods in the cluster, then ensure that you specify the necessary CIDR blocks. For more information, see Amazon EKS cluster endpoint access control in the /Amazon EKS User Guide/ .