Network.AWS.EKS.Types.VpcConfigResponse

Description

Documentation

data VpcConfigResponse

An object representing an Amazon EKS cluster VPC configuration response.

See: newVpcConfigResponse smart constructor.

Constructors

VpcConfigResponse'

Fields

securityGroupIds :: Maybe [Text]

The security groups associated with the cross-account elastic network interfaces that are used to allow communication between your nodes and the Kubernetes control plane.

endpointPrivateAccess :: Maybe Bool

This parameter indicates whether the Amazon EKS private API server endpoint is enabled. If the Amazon EKS private API server endpoint is enabled, Kubernetes API requests that originate from within your cluster's VPC use the private VPC endpoint instead of traversing the internet. If this value is disabled and you have nodes or Fargate pods in the cluster, then ensure that publicAccessCidrs includes the necessary CIDR blocks for communication with the nodes or Fargate pods. For more information, see Amazon EKS cluster endpoint access control in the /Amazon EKS User Guide/ .

publicAccessCidrs :: Maybe [Text]

The CIDR blocks that are allowed access to your cluster's public Kubernetes API server endpoint. Communication to the endpoint from addresses outside of the listed CIDR blocks is denied. The default value is 0.0.0.0/0. If you've disabled private endpoint access and you have nodes or Fargate pods in the cluster, then ensure that the necessary CIDR blocks are listed. For more information, see Amazon EKS cluster endpoint access control in the /Amazon EKS User Guide/ .

subnetIds :: Maybe [Text]

The subnets associated with your cluster.

vpcId :: Maybe Text

The VPC associated with your cluster.

clusterSecurityGroupId :: Maybe Text

The cluster security group that was created by Amazon EKS for the cluster. Managed node groups use this security group for control-plane-to-data-plane communication.

endpointPublicAccess :: Maybe Bool

This parameter indicates whether the Amazon EKS public API server endpoint is enabled. If the Amazon EKS public API server endpoint is disabled, your cluster's Kubernetes API server can only receive requests that originate from within the cluster VPC.

newVpcConfigResponse :: VpcConfigResponse

Create a value of VpcConfigResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:securityGroupIds:VpcConfigResponse', vpcConfigResponse_securityGroupIds - The security groups associated with the cross-account elastic network interfaces that are used to allow communication between your nodes and the Kubernetes control plane.

$sel:endpointPrivateAccess:VpcConfigResponse', vpcConfigResponse_endpointPrivateAccess - This parameter indicates whether the Amazon EKS private API server endpoint is enabled. If the Amazon EKS private API server endpoint is enabled, Kubernetes API requests that originate from within your cluster's VPC use the private VPC endpoint instead of traversing the internet. If this value is disabled and you have nodes or Fargate pods in the cluster, then ensure that publicAccessCidrs includes the necessary CIDR blocks for communication with the nodes or Fargate pods. For more information, see Amazon EKS cluster endpoint access control in the /Amazon EKS User Guide/ .

$sel:publicAccessCidrs:VpcConfigResponse', vpcConfigResponse_publicAccessCidrs - The CIDR blocks that are allowed access to your cluster's public Kubernetes API server endpoint. Communication to the endpoint from addresses outside of the listed CIDR blocks is denied. The default value is 0.0.0.0/0. If you've disabled private endpoint access and you have nodes or Fargate pods in the cluster, then ensure that the necessary CIDR blocks are listed. For more information, see Amazon EKS cluster endpoint access control in the /Amazon EKS User Guide/ .

$sel:subnetIds:VpcConfigResponse', vpcConfigResponse_subnetIds - The subnets associated with your cluster.

$sel:vpcId:VpcConfigResponse', vpcConfigResponse_vpcId - The VPC associated with your cluster.

$sel:clusterSecurityGroupId:VpcConfigResponse', vpcConfigResponse_clusterSecurityGroupId - The cluster security group that was created by Amazon EKS for the cluster. Managed node groups use this security group for control-plane-to-data-plane communication.

$sel:endpointPublicAccess:VpcConfigResponse', vpcConfigResponse_endpointPublicAccess - This parameter indicates whether the Amazon EKS public API server endpoint is enabled. If the Amazon EKS public API server endpoint is disabled, your cluster's Kubernetes API server can only receive requests that originate from within the cluster VPC.

vpcConfigResponse_securityGroupIds :: Lens' VpcConfigResponse (Maybe [Text])

The security groups associated with the cross-account elastic network interfaces that are used to allow communication between your nodes and the Kubernetes control plane.

vpcConfigResponse_endpointPrivateAccess :: Lens' VpcConfigResponse (Maybe Bool)

This parameter indicates whether the Amazon EKS private API server endpoint is enabled. If the Amazon EKS private API server endpoint is enabled, Kubernetes API requests that originate from within your cluster's VPC use the private VPC endpoint instead of traversing the internet. If this value is disabled and you have nodes or Fargate pods in the cluster, then ensure that publicAccessCidrs includes the necessary CIDR blocks for communication with the nodes or Fargate pods. For more information, see Amazon EKS cluster endpoint access control in the /Amazon EKS User Guide/ .

vpcConfigResponse_publicAccessCidrs :: Lens' VpcConfigResponse (Maybe [Text])

The CIDR blocks that are allowed access to your cluster's public Kubernetes API server endpoint. Communication to the endpoint from addresses outside of the listed CIDR blocks is denied. The default value is 0.0.0.0/0. If you've disabled private endpoint access and you have nodes or Fargate pods in the cluster, then ensure that the necessary CIDR blocks are listed. For more information, see Amazon EKS cluster endpoint access control in the /Amazon EKS User Guide/ .

vpcConfigResponse_subnetIds :: Lens' VpcConfigResponse (Maybe [Text])

The subnets associated with your cluster.

vpcConfigResponse_vpcId :: Lens' VpcConfigResponse (Maybe Text)

The VPC associated with your cluster.

vpcConfigResponse_clusterSecurityGroupId :: Lens' VpcConfigResponse (Maybe Text)

The cluster security group that was created by Amazon EKS for the cluster. Managed node groups use this security group for control-plane-to-data-plane communication.

vpcConfigResponse_endpointPublicAccess :: Lens' VpcConfigResponse (Maybe Bool)

This parameter indicates whether the Amazon EKS public API server endpoint is enabled. If the Amazon EKS public API server endpoint is disabled, your cluster's Kubernetes API server can only receive requests that originate from within the cluster VPC.