Network.AWS.EKS.UpdateNodegroupConfig

Description

Updates an Amazon EKS managed node group configuration. Your node group continues to function during the update. The response output includes an update ID that you can use to track the status of your node group update with the DescribeUpdate API operation. Currently you can update the Kubernetes labels for a node group or the scaling configuration.

Creating a Request

data UpdateNodegroupConfig

See: newUpdateNodegroupConfig smart constructor.

Constructors

UpdateNodegroupConfig'

Fields

scalingConfig :: Maybe NodegroupScalingConfig

The scaling configuration details for the Auto Scaling group after the update.

labels :: Maybe UpdateLabelsPayload

The Kubernetes labels to be applied to the nodes in the node group after the update.

updateConfig :: Maybe NodegroupUpdateConfig

The node group update configuration.

clientRequestToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

taints :: Maybe UpdateTaintsPayload

The Kubernetes taints to be applied to the nodes in the node group after the update.

clusterName :: Text

The name of the Amazon EKS cluster that the managed node group resides in.

nodegroupName :: Text

The name of the managed node group to update.

newUpdateNodegroupConfig

Arguments

:: Text
$sel:clusterName:UpdateNodegroupConfig'

-> Text
$sel:nodegroupName:UpdateNodegroupConfig'

-> UpdateNodegroupConfig

Create a value of UpdateNodegroupConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:scalingConfig:UpdateNodegroupConfig', updateNodegroupConfig_scalingConfig - The scaling configuration details for the Auto Scaling group after the update.

$sel:labels:UpdateNodegroupConfig', updateNodegroupConfig_labels - The Kubernetes labels to be applied to the nodes in the node group after the update.

$sel:updateConfig:UpdateNodegroupConfig', updateNodegroupConfig_updateConfig - The node group update configuration.

$sel:clientRequestToken:UpdateNodegroupConfig', updateNodegroupConfig_clientRequestToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

$sel:taints:UpdateNodegroupConfig', updateNodegroupConfig_taints - The Kubernetes taints to be applied to the nodes in the node group after the update.

$sel:clusterName:UpdateNodegroupConfig', updateNodegroupConfig_clusterName - The name of the Amazon EKS cluster that the managed node group resides in.

$sel:nodegroupName:UpdateNodegroupConfig', updateNodegroupConfig_nodegroupName - The name of the managed node group to update.

Request Lenses

updateNodegroupConfig_scalingConfig :: Lens' UpdateNodegroupConfig (Maybe NodegroupScalingConfig)

The scaling configuration details for the Auto Scaling group after the update.

updateNodegroupConfig_labels :: Lens' UpdateNodegroupConfig (Maybe UpdateLabelsPayload)

The Kubernetes labels to be applied to the nodes in the node group after the update.

updateNodegroupConfig_updateConfig :: Lens' UpdateNodegroupConfig (Maybe NodegroupUpdateConfig)

The node group update configuration.

updateNodegroupConfig_clientRequestToken :: Lens' UpdateNodegroupConfig (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

updateNodegroupConfig_taints :: Lens' UpdateNodegroupConfig (Maybe UpdateTaintsPayload)

The Kubernetes taints to be applied to the nodes in the node group after the update.

updateNodegroupConfig_clusterName :: Lens' UpdateNodegroupConfig Text

The name of the Amazon EKS cluster that the managed node group resides in.

updateNodegroupConfig_nodegroupName :: Lens' UpdateNodegroupConfig Text

The name of the managed node group to update.

Destructuring the Response

data UpdateNodegroupConfigResponse

See: newUpdateNodegroupConfigResponse smart constructor.

Constructors

UpdateNodegroupConfigResponse'

Fields

update :: Maybe Update

httpStatus :: Int

The response's http status code.

newUpdateNodegroupConfigResponse

Arguments

:: Int
$sel:httpStatus:UpdateNodegroupConfigResponse'

-> UpdateNodegroupConfigResponse

Create a value of UpdateNodegroupConfigResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:update:UpdateNodegroupConfigResponse', updateNodegroupConfigResponse_update - Undocumented member.

$sel:httpStatus:UpdateNodegroupConfigResponse', updateNodegroupConfigResponse_httpStatus - The response's http status code.

Response Lenses

updateNodegroupConfigResponse_update :: Lens' UpdateNodegroupConfigResponse (Maybe Update)

Undocumented member.

updateNodegroupConfigResponse_httpStatus :: Lens' UpdateNodegroupConfigResponse Int

The response's http status code.