Network.AWS.EKS.UpdateNodegroupVersion

Description

Updates the Kubernetes version or AMI version of an Amazon EKS managed node group.

You can update a node group using a launch template only if the node group was originally deployed with a launch template. If you need to update a custom AMI in a node group that was deployed with a launch template, then update your custom AMI, specify the new ID in a new version of the launch template, and then update the node group to the new version of the launch template.

If you update without a launch template, then you can update to the latest available AMI version of a node group's current Kubernetes version by not specifying a Kubernetes version in the request. You can update to the latest AMI version of your cluster's current Kubernetes version by specifying your cluster's Kubernetes version in the request. For more information, see Amazon EKS optimized Amazon Linux 2 AMI versions in the Amazon EKS User Guide.

You cannot roll back a node group to an earlier Kubernetes version or AMI version.

When a node in a managed node group is terminated due to a scaling action or update, the pods in that node are drained first. Amazon EKS attempts to drain the nodes gracefully and will fail if it is unable to do so. You can force the update if Amazon EKS is unable to drain the nodes as a result of a pod disruption budget issue.

Creating a Request

data UpdateNodegroupVersion

See: newUpdateNodegroupVersion smart constructor.

Constructors

UpdateNodegroupVersion'

Fields

force :: Maybe Bool

Force the update if the existing node group's pods are unable to be drained due to a pod disruption budget issue. If an update fails because pods could not be drained, you can force the update after it fails to terminate the old node whether or not any pods are running on the node.

releaseVersion :: Maybe Text

The AMI version of the Amazon EKS optimized AMI to use for the update. By default, the latest available AMI version for the node group's Kubernetes version is used. For more information, see Amazon EKS optimized Amazon Linux 2 AMI versions in the Amazon EKS User Guide. If you specify launchTemplate, and your launch template uses a custom AMI, then don't specify releaseVersion, or the node group update will fail. For more information about using launch templates with Amazon EKS, see Launch template support in the Amazon EKS User Guide.

version :: Maybe Text

The Kubernetes version to update to. If no version is specified, then the Kubernetes version of the node group does not change. You can specify the Kubernetes version of the cluster to update the node group to the latest AMI version of the cluster's Kubernetes version. If you specify launchTemplate, and your launch template uses a custom AMI, then don't specify version, or the node group update will fail. For more information about using launch templates with Amazon EKS, see Launch template support in the Amazon EKS User Guide.

launchTemplate :: Maybe LaunchTemplateSpecification

An object representing a node group's launch template specification. You can only update a node group using a launch template if the node group was originally deployed with a launch template.

clientRequestToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

clusterName :: Text

The name of the Amazon EKS cluster that is associated with the managed node group to update.

nodegroupName :: Text

The name of the managed node group to update.

newUpdateNodegroupVersion

Arguments

:: Text
$sel:clusterName:UpdateNodegroupVersion'

-> Text
$sel:nodegroupName:UpdateNodegroupVersion'

-> UpdateNodegroupVersion

Create a value of UpdateNodegroupVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:force:UpdateNodegroupVersion', updateNodegroupVersion_force - Force the update if the existing node group's pods are unable to be drained due to a pod disruption budget issue. If an update fails because pods could not be drained, you can force the update after it fails to terminate the old node whether or not any pods are running on the node.

$sel:releaseVersion:UpdateNodegroupVersion', updateNodegroupVersion_releaseVersion - The AMI version of the Amazon EKS optimized AMI to use for the update. By default, the latest available AMI version for the node group's Kubernetes version is used. For more information, see Amazon EKS optimized Amazon Linux 2 AMI versions in the Amazon EKS User Guide. If you specify launchTemplate, and your launch template uses a custom AMI, then don't specify releaseVersion, or the node group update will fail. For more information about using launch templates with Amazon EKS, see Launch template support in the Amazon EKS User Guide.

$sel:version:UpdateNodegroupVersion', updateNodegroupVersion_version - The Kubernetes version to update to. If no version is specified, then the Kubernetes version of the node group does not change. You can specify the Kubernetes version of the cluster to update the node group to the latest AMI version of the cluster's Kubernetes version. If you specify launchTemplate, and your launch template uses a custom AMI, then don't specify version, or the node group update will fail. For more information about using launch templates with Amazon EKS, see Launch template support in the Amazon EKS User Guide.

$sel:launchTemplate:UpdateNodegroupVersion', updateNodegroupVersion_launchTemplate - An object representing a node group's launch template specification. You can only update a node group using a launch template if the node group was originally deployed with a launch template.

$sel:clientRequestToken:UpdateNodegroupVersion', updateNodegroupVersion_clientRequestToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

$sel:clusterName:UpdateNodegroupVersion', updateNodegroupVersion_clusterName - The name of the Amazon EKS cluster that is associated with the managed node group to update.

$sel:nodegroupName:UpdateNodegroupVersion', updateNodegroupVersion_nodegroupName - The name of the managed node group to update.

Request Lenses

updateNodegroupVersion_force :: Lens' UpdateNodegroupVersion (Maybe Bool)

Force the update if the existing node group's pods are unable to be drained due to a pod disruption budget issue. If an update fails because pods could not be drained, you can force the update after it fails to terminate the old node whether or not any pods are running on the node.

updateNodegroupVersion_releaseVersion :: Lens' UpdateNodegroupVersion (Maybe Text)

The AMI version of the Amazon EKS optimized AMI to use for the update. By default, the latest available AMI version for the node group's Kubernetes version is used. For more information, see Amazon EKS optimized Amazon Linux 2 AMI versions in the Amazon EKS User Guide. If you specify launchTemplate, and your launch template uses a custom AMI, then don't specify releaseVersion, or the node group update will fail. For more information about using launch templates with Amazon EKS, see Launch template support in the Amazon EKS User Guide.

updateNodegroupVersion_version :: Lens' UpdateNodegroupVersion (Maybe Text)

The Kubernetes version to update to. If no version is specified, then the Kubernetes version of the node group does not change. You can specify the Kubernetes version of the cluster to update the node group to the latest AMI version of the cluster's Kubernetes version. If you specify launchTemplate, and your launch template uses a custom AMI, then don't specify version, or the node group update will fail. For more information about using launch templates with Amazon EKS, see Launch template support in the Amazon EKS User Guide.

updateNodegroupVersion_launchTemplate :: Lens' UpdateNodegroupVersion (Maybe LaunchTemplateSpecification)

An object representing a node group's launch template specification. You can only update a node group using a launch template if the node group was originally deployed with a launch template.

updateNodegroupVersion_clientRequestToken :: Lens' UpdateNodegroupVersion (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

updateNodegroupVersion_clusterName :: Lens' UpdateNodegroupVersion Text

The name of the Amazon EKS cluster that is associated with the managed node group to update.

updateNodegroupVersion_nodegroupName :: Lens' UpdateNodegroupVersion Text

The name of the managed node group to update.

Destructuring the Response

data UpdateNodegroupVersionResponse

See: newUpdateNodegroupVersionResponse smart constructor.

Constructors

UpdateNodegroupVersionResponse'

Fields

update :: Maybe Update

httpStatus :: Int

The response's http status code.

newUpdateNodegroupVersionResponse

Arguments

:: Int
$sel:httpStatus:UpdateNodegroupVersionResponse'

-> UpdateNodegroupVersionResponse

Create a value of UpdateNodegroupVersionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:update:UpdateNodegroupVersionResponse', updateNodegroupVersionResponse_update - Undocumented member.

$sel:httpStatus:UpdateNodegroupVersionResponse', updateNodegroupVersionResponse_httpStatus - The response's http status code.

Response Lenses

updateNodegroupVersionResponse_update :: Lens' UpdateNodegroupVersionResponse (Maybe Update)

Undocumented member.

updateNodegroupVersionResponse_httpStatus :: Lens' UpdateNodegroupVersionResponse Int

The response's http status code.