Network.AWS.ELB

Description

Elastic Load Balancing helps you improve the availability and scalability of your application running on Amazon Elastic Cloud Compute (Amazon EC2).

Actions

ApplySecurityGroupsToLoadBalancer

data ApplySecurityGroupsToLoadBalancer

Associates one or more security groups with your load balancer in VPC. The provided security group IDs will override any currently applied security groups.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_ApplySecurityGroupsToLoadBalancer.html

Constructors

ApplySecurityGroupsToLoadBalancer

Fields

asgtlbLoadBalancerName :: !Text

The name associated with the load balancer. The name must be unique within the client AWS account.

asgtlbSecurityGroups :: Members Text

A list of security group IDs to associate with your load balancer in VPC. The security group IDs must be provided as the ID and not the security group name (For example, sg-1234).

data ApplySecurityGroupsToLoadBalancerResponse

Constructors

ApplySecurityGroupsToLoadBalancerResponse

Fields

asgtlbrResponseMetadata :: !Text

asgtlbrApplySecurityGroupsToLoadBalancerResult :: !ApplySecurityGroupsToLoadBalancerResult

AttachLoadBalancerToSubnets

data AttachLoadBalancerToSubnets

Adds one or more subnets to the set of configured subnets in the VPC for the load balancer. The Loadbalancers evenly distribute requests across all of the registered subnets.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_AttachLoadBalancerToSubnets.html

Constructors

AttachLoadBalancerToSubnets

Fields

albtsLoadBalancerName :: !Text

The name associated with the load balancer. The name must be unique within the client AWS account.

albtsSubnets :: Members Text

A list of subnet IDs to add for the load balancer.

data AttachLoadBalancerToSubnetsResponse

Constructors

AttachLoadBalancerToSubnetsResponse

Fields

albtsrResponseMetadata :: !Text

albtsrAttachLoadBalancerToSubnetsResult :: !AttachLoadBalancerToSubnetsResult

ConfigureHealthCheck

data ConfigureHealthCheck

Enables the client to define an application healthcheck for the instances.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_ConfigureHealthCheck.html

Constructors

ConfigureHealthCheck

Fields

chcHealthCheck :: !HealthCheck

A structure containing the configuration information for the new healthcheck.

chcLoadBalancerName :: !Text

The mnemonic name associated with the load balancer. This name must be unique within the client AWS account.

data ConfigureHealthCheckResponse

Constructors

ConfigureHealthCheckResponse

Fields

chcrResponseMetadata :: !Text

chcrConfigureHealthCheckResult :: !ConfigureHealthCheckResult

CreateAppCookieStickinessPolicy

data CreateAppCookieStickinessPolicy

Generates a stickiness policy with sticky session lifetimes that follow that of an application-generated cookie. This policy can be associated only with HTTP/HTTPS listeners. This policy is similar to the policy created by CreatelbCookieStickinessPolicy, except that the lifetime of the special Elastic Load Balancing cookie follows the lifetime of the application-generated cookie specified in the policy configuration. The load balancer only inserts a new stickiness cookie when the application response includes a new application cookie. If the application cookie is explicitly removed or expires, the session stops being sticky until a new application cookie is issued. Note An application client must receive and send two cookies: the application-generated cookie and the special Elastic Load Balancing cookie named AWSELB. This is the default behavior for many common web browsers. For information on using CreateAppCookieStickinessPolicy, see Using the Query API in the Enabling Application-Controlled Sesssion Stickiness section of the Elastic Load Balancing Developer Guide.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_CreateAppCookieStickinessPolicy.html

Constructors

CreateAppCookieStickinessPolicy

Fields

cacspCookieName :: !Text

Name of the application cookie used for stickiness.

cacspLoadBalancerName :: !Text

The name associated with the load balancer. The name must be unique within the client AWS account.

cacspPolicyName :: !Text

The name of the policy being created. The name must be unique within the set of policies for this load balancer.

data CreateAppCookieStickinessPolicyResponse

Constructors

CreateAppCookieStickinessPolicyResponse

Fields

cacsprResponseMetadata :: !Text

cacsprCreateAppCookieStickinessPolicyResult :: !CreateAppCookieStickinessPolicyResult

CreatelbCookieStickinessPolicy

data CreatelbCookieStickinessPolicy

Generates a stickiness policy with sticky session lifetimes controlled by the lifetime of the browser (user-agent) or a specified expiration period. This policy can be associated only with HTTP/HTTPS listeners. When a load balancer implements this policy, the load balancer uses a special cookie to track the backend server instance for each request. When the load balancer receives a request, it first checks to see if this cookie is present in the request. If so, the load balancer sends the request to the application server specified in the cookie. If not, the load balancer sends the request to a server that is chosen based on the existing load balancing algorithm. A cookie is inserted into the response for binding subsequent requests from the same user to that server. The validity of the cookie is based on the cookie expiration time, which is specified in the policy configuration. For information on using CreatelbCookieStickinessPolicy, see Using the Query API in the Enabling Duration-Based Sesssion Stickiness section of the Elastic Load Balancing Developer Guide.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_CreatelbCookieStickinessPolicy.html

Constructors

CreatelbCookieStickinessPolicy

Fields

clbcspCookieExpirationPeriod :: Maybe Integer

The time period in seconds after which the cookie should be considered stale. Not specifying this parameter indicates that the sticky session will last for the duration of the browser session.

clbcspLoadBalancerName :: !Text

The name associated with the load balancer. The name must be unique within the client AWS account.

clbcspPolicyName :: !Text

The name of the policy being created. The name must be unique within the set of policies for this load balancer.

data CreatelbCookieStickinessPolicyResponse

Constructors

CreatelbCookieStickinessPolicyResponse

Fields

clbcsprResponseMetadata :: !Text

clbcsprCreatelbCookieStickinessPolicyResult :: !CreatelbCookieStickinessPolicyResult

CreateLoadBalancer

data CreateLoadBalancer

Creates a new load balancer. After the call has completed successfully, a new load balancer is created; however, it will not be usable until at least one instance has been registered. When the load balancer creation is completed, the client can check whether or not it is usable by using the DescribeInstanceHealth action. The load balancer is usable as soon as any registered instance is InService. Note Currently, the client's quota of load balancers is limited to ten per Region. Note Load balancer DNS names vary depending on the Region they're created in. For load balancers created in the United States, the DNS name ends with: For load balancers created in the EU (Ireland) Region, the DNS name ends with: For information on using CreateLoadBalancer to create a new load balancer in Amazon EC2, go to Using Query API section in the Creating a Load Balancer With SSL Cipher Settings and Back-end Authentication topic of the Elastic Load Balancing Developer Guide. For information on using CreateLoadBalancer to create a new load balancer in Amazon VPC, see Using the Query API in the Creating a Basic Load Balancer in Amazon VPC section in the Elastic Load Balancing Developer Guide.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_CreateLoadBalancer.html

Constructors

CreateLoadBalancer

Fields

clbAvailabilityZones :: Members Text

A list of Availability Zones.

clbListeners :: Members Listener

A list of the following tuples: LoadBalancerPort, InstancePort, and Protocol.

clbLoadBalancerName :: !Text

The name associated with the load balancer. The name must be unique within your set of load balancers.

clbScheme :: Maybe Text

The type of a load balancer.

clbSecurityGroups :: Members Text

The security groups assigned to your load balancer within your VPC.

clbSubnets :: Members Text

A list of subnet IDs in your VPC to attach to your load balancer.

data CreateLoadBalancerResponse

Constructors

CreateLoadBalancerResponse

Fields

clbrResponseMetadata :: !Text

clbrCreateLoadBalancerResult :: !CreateLoadBalancerResult

CreateLoadBalancerListeners

data CreateLoadBalancerListeners

Creates one or more listeners on a load balancer for the specified port. If a listener with the given port does not already exist, it will be created; otherwise, the properties of the new listener must match the properties of the existing listener.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_CreateLoadBalancerListeners.html

Constructors

CreateLoadBalancerListeners

Fields

clblListeners :: Members Listener

A list of LoadBalancerPort, InstancePort, Protocol, and SSLCertificateId items.

clblLoadBalancerName :: !Text

The name of the load balancer.

data CreateLoadBalancerListenersResponse

Constructors

CreateLoadBalancerListenersResponse

Fields

clblrResponseMetadata :: !Text

clblrCreateLoadBalancerListenersResult :: !CreateLoadBalancerListenersResult

CreateLoadBalancerPolicy

data CreateLoadBalancerPolicy

Creates a new policy that contains the necessary attributes depending on the policy type. Policies are settings that are saved for your load balancer and that can be applied to the front-end listener, or the back-end application server, depending on your policy type.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_CreateLoadBalancerPolicy.html

Constructors

CreateLoadBalancerPolicy

Fields

clbpLoadBalancerName :: !Text

The name associated with the LoadBalancer for which the policy is being created. This name must be unique within the client AWS account.

clbpPolicyAttributes :: Members PolicyAttribute

A list of attributes associated with the policy being created.

clbpPolicyName :: !Text

The name of the load balancer policy being created. The name must be unique within the set of policies for this load balancer.

clbpPolicyTypeName :: !Text

The name of the base policy type being used to create this policy. To get the list of policy types, use the DescribeLoadBalancerPolicyTypes action.

data CreateLoadBalancerPolicyResponse

Constructors

CreateLoadBalancerPolicyResponse

Fields

clbprResponseMetadata :: !Text

clbprCreateLoadBalancerPolicyResult :: !CreateLoadBalancerPolicyResult

DeleteLoadBalancer

data DeleteLoadBalancer

Deletes the specified load balancer. If attempting to recreate the load balancer, the client must reconfigure all the settings. The DNS name associated with a deleted load balancer will no longer be usable. Once deleted, the name and associated DNS record of the load balancer no longer exist and traffic sent to any of its IP addresses will no longer be delivered to client instances. The client will not receive the same DNS name even if a new load balancer with same load balancer name is created. To successfully call this API, the client must provide the same account credentials as were used to create the load balancer. Note By design, if the load balancer does not exist or has already been deleted, DeleteLoadBalancer still succeeds.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_DeleteLoadBalancer.html

Constructors

DeleteLoadBalancer

Fields

dlbLoadBalancerName :: !Text

The name associated with the load balancer. The name must be unique within the client AWS account.

data DeleteLoadBalancerResponse

Constructors

DeleteLoadBalancerResponse

Fields

dlbrResponseMetadata :: !Text

dlbrDeleteLoadBalancerResult :: !DeleteLoadBalancerResult

DeleteLoadBalancerListeners

data DeleteLoadBalancerListeners

Deletes listeners from the load balancer for the specified port.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_DeleteLoadBalancerListeners.html

Constructors

DeleteLoadBalancerListeners

Fields

dlblLoadBalancerName :: !Text

The mnemonic name associated with the load balancer.

dlblLoadBalancerPorts :: Members Integer

The client port number(s) of the load balancer listener(s) to be removed.

data DeleteLoadBalancerListenersResponse

Constructors

DeleteLoadBalancerListenersResponse

Fields

dlblrResponseMetadata :: !Text

dlblrDeleteLoadBalancerListenersResult :: !DeleteLoadBalancerListenersResult

DeleteLoadBalancerPolicy

data DeleteLoadBalancerPolicy

Deletes a policy from the load balancer. The specified policy must not be enabled for any listeners.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_DeleteLoadBalancerPolicy.html

Constructors

DeleteLoadBalancerPolicy

Fields

dlbpLoadBalancerName :: !Text

The mnemonic name associated with the load balancer. The name must be unique within your AWS account.

dlbpPolicyName :: !Text

The mnemonic name for the policy being deleted.

data DeleteLoadBalancerPolicyResponse

Constructors

DeleteLoadBalancerPolicyResponse

Fields

dlbprResponseMetadata :: !Text

dlbprDeleteLoadBalancerPolicyResult :: !DeleteLoadBalancerPolicyResult

DeregisterInstancesFromLoadBalancer

data DeregisterInstancesFromLoadBalancer

Deregisters instances from the load balancer. Once the instance is deregistered, it will stop receiving traffic from the load balancer. In order to successfully call this API, the same account credentials as those used to create the load balancer must be provided.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_DeregisterInstancesFromLoadBalancer.html

Constructors

DeregisterInstancesFromLoadBalancer

Fields

diflbInstances :: Members Instance

A list of EC2 instance IDs consisting of all instances to be deregistered.

diflbLoadBalancerName :: !Text

The name associated with the load balancer. The name must be unique within the client AWS account.

data DeregisterInstancesFromLoadBalancerResponse

Constructors

DeregisterInstancesFromLoadBalancerResponse

Fields

diflbrResponseMetadata :: !Text

diflbrDeregisterInstancesFromLoadBalancerResult :: !DeregisterInstancesFromLoadBalancerResult

DescribeInstanceHealth

data DescribeInstanceHealth

Returns the current state of the instances of the specified load balancer. If no instances are specified, the state of all the instances for the load balancer is returned. Note The client must have created the specified input load balancer in order to retrieve this information; the client must provide the same account credentials as those that were used to create the load balancer.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_DescribeInstanceHealth.html

Constructors

DescribeInstanceHealth

Fields

dihInstances :: Members Instance

A list of instance IDs whose states are being queried.

dihLoadBalancerName :: !Text

The name associated with the load balancer. The name must be unique within the client AWS account.

data DescribeInstanceHealthResponse

Constructors

DescribeInstanceHealthResponse

Fields

dihrResponseMetadata :: !Text

dihrDescribeInstanceHealthResult :: !DescribeInstanceHealthResult

DescribeLoadBalancerPolicies

data DescribeLoadBalancerPolicies

Returns detailed descriptions of the policies. If you specify a load balancer name, the operation returns either the descriptions of the specified policies, or descriptions of all the policies created for the load balancer. If you don't specify a load balancer name, the operation returns descriptions of the specified sample policies, or descriptions of all the sample policies. The names of the sample policies have the ELBSample- prefix.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_DescribeLoadBalancerPolicies.html

Constructors

DescribeLoadBalancerPolicies

Fields

dlbqLoadBalancerName :: Maybe Text

The mnemonic name associated with the load balancer. If no name is specified, the operation returns the attributes of either all the sample policies pre-defined by Elastic Load Balancing or the specified sample polices.

dlbqPolicyNames :: Members Text

The names of LoadBalancer policies you've created or Elastic Load Balancing sample policy names.

data DescribeLoadBalancerPoliciesResponse

Constructors

DescribeLoadBalancerPoliciesResponse

Fields

dlbpsResponseMetadata :: !Text

dlbpsDescribeLoadBalancerPoliciesResult :: !DescribeLoadBalancerPoliciesResult

DescribeLoadBalancerPolicyTypes

data DescribeLoadBalancerPolicyTypes

Returns meta-information on the specified load balancer policies defined by the Elastic Load Balancing service. The policy types that are returned from this action can be used in a CreateLoadBalancerPolicy action to instantiate specific policy configurations that will be applied to a load balancer.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_DescribeLoadBalancerPolicyTypes.html

Constructors

DescribeLoadBalancerPolicyTypes

Fields

dlbptPolicyTypeNames :: Members Text

Specifies the name of the policy types. If no names are specified, returns the description of all the policy types defined by Elastic Load Balancing service.

data DescribeLoadBalancerPolicyTypesResponse

Constructors

DescribeLoadBalancerPolicyTypesResponse

Fields

dlbptrResponseMetadata :: !Text

dlbptrDescribeLoadBalancerPolicyTypesResult :: !DescribeLoadBalancerPolicyTypesResult

DescribeLoadBalancers

data DescribeLoadBalancers

Returns detailed configuration information for the specified load balancers. If no load balancers are specified, the operation returns configuration information for all load balancers created by the caller. Note The client must have created the specified input load balancers in order to retrieve this information; the client must provide the same account credentials as those that were used to create the load balancer.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_DescribeLoadBalancers.html

Constructors

DescribeLoadBalancers

Fields

dlbLoadBalancerNames :: Members Text

A list of names associated with the load balancers at creation time.

dlbMarker :: Maybe Text

An optional parameter reserved for future use.

data DescribeLoadBalancersResponse

Constructors

DescribeLoadBalancersResponse

Fields

dlbsResponseMetadata :: !Text

dlbsDescribeLoadBalancersResult :: !DescribeLoadBalancersResult

DetachLoadBalancerFromSubnets

data DetachLoadBalancerFromSubnets

Removes subnets from the set of configured subnets in the VPC for the load balancer. After a subnet is removed all of the EndPoints registered with the load balancer that are in the removed subnet will go into the OutOfService state. When a subnet is removed, the load balancer will balance the traffic among the remaining routable subnets for the load balancer.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_DetachLoadBalancerFromSubnets.html

Constructors

DetachLoadBalancerFromSubnets

Fields

dlbfsLoadBalancerName :: !Text

The name associated with the load balancer to be detached. The name must be unique within the client AWS account.

dlbfsSubnets :: Members Text

A list of subnet IDs to remove from the set of configured subnets for the load balancer.

data DetachLoadBalancerFromSubnetsResponse

Constructors

DetachLoadBalancerFromSubnetsResponse

Fields

dlbfsrResponseMetadata :: !Text

dlbfsrDetachLoadBalancerFromSubnetsResult :: !DetachLoadBalancerFromSubnetsResult

DisableAvailabilityZonesForLoadBalancer

data DisableAvailabilityZonesForLoadBalancer

Removes the specified EC2 Availability Zones from the set of configured Availability Zones for the load balancer. There must be at least one Availability Zone registered with a load balancer at all times. A client cannot remove all the Availability Zones from a load balancer. Once an Availability Zone is removed, all the instances registered with the load balancer that are in the removed Availability Zone go into the OutOfService state. Upon Availability Zone removal, the load balancer attempts to equally balance the traffic among its remaining usable Availability Zones. Trying to remove an Availability Zone that was not associated with the load balancer does nothing. Note In order for this call to be successful, the client must have created the load balancer. The client must provide the same account credentials as those that were used to create the load balancer.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_DisableAvailabilityZonesForLoadBalancer.html

Constructors

DisableAvailabilityZonesForLoadBalancer

Fields

dazflbAvailabilityZones :: Members Text

A list of Availability Zones to be removed from the load balancer.

dazflbLoadBalancerName :: !Text

The name associated with the load balancer. The name must be unique within the client AWS account.

data DisableAvailabilityZonesForLoadBalancerResponse

Constructors

DisableAvailabilityZonesForLoadBalancerResponse

Fields

dazflbrResponseMetadata :: !Text

dazflbrDisableAvailabilityZonesForLoadBalancerResult :: !DisableAvailabilityZonesForLoadBalancerResult

EnableAvailabilityZonesForLoadBalancer

data EnableAvailabilityZonesForLoadBalancer

Adds one or more EC2 Availability Zones to the load balancer. The load balancer evenly distributes requests across all its registered Availability Zones that contain instances. As a result, the client must ensure that its load balancer is appropriately scaled for each registered Availability Zone. Note The new EC2 Availability Zones to be added must be in the same EC2 Region as the Availability Zones for which the load balancer was created.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_EnableAvailabilityZonesForLoadBalancer.html

Constructors

EnableAvailabilityZonesForLoadBalancer

Fields

eazflbAvailabilityZones :: Members Text

A list of new Availability Zones for the load balancer. Each Availability Zone must be in the same Region as the load balancer.

eazflbLoadBalancerName :: !Text

The name associated with the load balancer. The name must be unique within the client AWS account.

data EnableAvailabilityZonesForLoadBalancerResponse

Constructors

EnableAvailabilityZonesForLoadBalancerResponse

Fields

eazflbrResponseMetadata :: !Text

eazflbrEnableAvailabilityZonesForLoadBalancerResult :: !EnableAvailabilityZonesForLoadBalancerResult

RegisterInstancesWithLoadBalancer

data RegisterInstancesWithLoadBalancer

Adds new instances to the load balancer. Once the instance is registered, it starts receiving traffic and requests from the load balancer. Any instance that is not in any of the Availability Zones registered for the load balancer will be moved to the OutOfService state. It will move to the InService state when the Availability Zone is added to the load balancer. Note In order for this call to be successful, the client must have created the load balancer. The client must provide the same account credentials as those that were used to create the load balancer. Note Completion of this API does not guarantee that operation has completed. Rather, it means that the request has been registered and the changes will happen shortly.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_RegisterInstancesWithLoadBalancer.html

Constructors

RegisterInstancesWithLoadBalancer

Fields

riwlbInstances :: Members Instance

A list of instance IDs that should be registered with the load balancer.

riwlbLoadBalancerName :: !Text

The name associated with the load balancer. The name must be unique within the client AWS account.

data RegisterInstancesWithLoadBalancerResponse

Constructors

RegisterInstancesWithLoadBalancerResponse

Fields

riwlbrResponseMetadata :: !Text

riwlbrRegisterInstancesWithLoadBalancerResult :: !RegisterInstancesWithLoadBalancerResult

SetLoadBalancerListenerSSLCertificate

data SetLoadBalancerListenerSSLCertificate

Sets the certificate that terminates the specified listener's SSL connections. The specified certificate replaces any prior certificate that was used on the same load balancer and port. For information on using SetLoadBalancerListenerSSLCertificate, see Using the Query API in the Updating an SSL Certificate for a Load Balancer section in of the Elastic Load Balancing Developer Guide.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_SetLoadBalancerListenerSSLCertificate.html

Constructors

SetLoadBalancerListenerSSLCertificate

Fields

slblsslcLoadBalancerName :: !Text

The name of the the load balancer.

slblsslcLoadBalancerPort :: !Integer

The port that uses the specified SSL certificate.

slblsslcSSLCertificateId :: !Text

The ID of the SSL certificate chain to use. For more information on SSL certificates, see Managing Server Certificates in the AWS Identity and Access Management documentation.

data SetLoadBalancerListenerSSLCertificateResponse

Constructors

SetLoadBalancerListenerSSLCertificateResponse

Fields

slblsslcrResponseMetadata :: !Text

slblsslcrSetLoadBalancerListenerSSLCertificateResult :: !SetLoadBalancerListenerSSLCertificateResult

SetLoadBalancerPoliciesForBackendServer

data SetLoadBalancerPoliciesForBackendServer

Replaces the current set of policies associated with a port on which the back-end server is listening with a new set of policies. After the policies have been created using CreateLoadBalancerPolicy, they can be applied here as a list. At this time, only the back-end server authentication policy type can be applied to the back-end ports; this policy type is composed of multiple public key policies.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_SetLoadBalancerPoliciesForBackendServer.html

Constructors

SetLoadBalancerPoliciesForBackendServer

Fields

slbpfbsInstancePort :: !Integer

The port number associated with the back-end server.

slbpfbsLoadBalancerName :: !Text

The mnemonic name associated with the load balancer. This name must be unique within the client AWS account.

slbpfbsPolicyNames :: Members Text

List of policy names to be set. If the list is empty, then all current polices are removed from the back-end server.

data SetLoadBalancerPoliciesForBackendServerResponse

Constructors

SetLoadBalancerPoliciesForBackendServerResponse

Fields

slbpfbsrResponseMetadata :: !Text

slbpfbsrSetLoadBalancerPoliciesForBackendServerResult :: !SetLoadBalancerPoliciesForBackendServerResult

SetLoadBalancerPoliciesOfListener

data SetLoadBalancerPoliciesOfListener

Associates, updates, or disables a policy with a listener on the load balancer. You can associate multiple policies with a listener.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_SetLoadBalancerPoliciesOfListener.html

Constructors

SetLoadBalancerPoliciesOfListener

Fields

slbpolLoadBalancerName :: !Text

The name associated with the load balancer. The name must be unique within the client AWS account.

slbpolLoadBalancerPort :: !Integer

The external port of the load balancer with which this policy applies to.

slbpolPolicyNames :: Members Text

List of policies to be associated with the listener. Currently this list can have at most one policy. If the list is empty, the current policy is removed from the listener.

data SetLoadBalancerPoliciesOfListenerResponse

Constructors

SetLoadBalancerPoliciesOfListenerResponse

Fields

slbpolrResponseMetadata :: !Text

slbpolrSetLoadBalancerPoliciesOfListenerResult :: !SetLoadBalancerPoliciesOfListenerResult

Data Types

elb :: Service

Currently supported version of the ELB service.

data ELBError

Constructors

ELBError

Fields

cweCode :: !Text

cweMessage :: !Text

data AppCookieStickinessPolicy

The AppCookieStickinessPolicy data type.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_AppCookieStickinessPolicy.html

Constructors

AppCookieStickinessPolicy

Fields

acspCookieName :: Maybe Text

The name of the application cookie used for stickiness.

acspPolicyName :: Maybe Text

The mnemonic name for the policy being created. The name must be unique within a set of policies for this load balancer.

data ApplySecurityGroupsToLoadBalancerResult

The out for the ApplySecurityGroupsToLoadBalancer action.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_ApplySecurityGroupsToLoadBalancerResult.html

Constructors

ApplySecurityGroupsToLoadBalancerResult

Fields

asgtlbrSecurityGroups :: Maybe Text

A list of security group IDs associated with your load balancer.

data AttachLoadBalancerToSubnetsResult

The output for the AttachLoadBalancerToSubnets action.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_AttachLoadBalancerToSubnetsResult.html

Constructors

AttachLoadBalancerToSubnetsResult

Fields

albtsrSubnets :: Maybe Text

A list of subnet IDs added for the load balancer.

data BackendServerDescription

This data type is used as a response element in the DescribeLoadBalancers action to describe the configuration of the back-end server.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_BackendServerDescription.html

Constructors

BackendServerDescription

Fields

bsdInstancePort :: Maybe Integer

Provides the port on which the back-end server is listening.

bsdPolicyNames :: Maybe Text

Provides a list of policy names enabled for the back-end server.

data ConfigureHealthCheckResult

The output for the ConfigureHealthCheck action.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_ConfigureHealthCheckResult.html

Constructors

ConfigureHealthCheckResult

Fields

chcrHealthCheck :: Maybe HealthCheck

The updated healthcheck for the instances.

data CreateAppCookieStickinessPolicyResult

The output for the CreateAppCookieStickinessPolicy action.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_CreateAppCookieStickinessPolicyResult.html

Constructors

CreateAppCookieStickinessPolicyResult

data CreatelbCookieStickinessPolicyResult

The output for the CreatelbCookieStickinessPolicy action.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_CreatelbCookieStickinessPolicyResult.html

Constructors

CreatelbCookieStickinessPolicyResult

data CreateLoadBalancerListenersResult

The output for the CreateLoadBalancerListeners action.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_CreateLoadBalancerListenersResult.html

Constructors

CreateLoadBalancerListenersResult

data CreateLoadBalancerPolicyResult

The output for the CreateLoadBalancerPolicy action.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_CreateLoadBalancerPolicyResult.html

Constructors

CreateLoadBalancerPolicyResult

data CreateLoadBalancerResult

The output for the CreateLoadBalancer action.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_CreateLoadBalancerResult.html

Constructors

CreateLoadBalancerResult

Fields

clbrDNSName :: Maybe Text

The DNS name for the load balancer.

data DeleteLoadBalancerListenersResult

The output for the DeleteLoadBalancerListeners action.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_DeleteLoadBalancerListenersResult.html

Constructors

DeleteLoadBalancerListenersResult

data DeleteLoadBalancerPolicyResult

The output for the DeleteLoadBalancerPolicy action.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_DeleteLoadBalancerPolicyResult.html

Constructors

DeleteLoadBalancerPolicyResult

data DeleteLoadBalancerResult

The output for the DeleteLoadBalancer action.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_DeleteLoadBalancerResult.html

Constructors

DeleteLoadBalancerResult

data DeregisterInstancesFromLoadBalancerResult

The output for the DeregisterInstancesFromLoadBalancer action.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_DeregisterInstancesFromLoadBalancerResult.html

Constructors

DeregisterInstancesFromLoadBalancerResult

Fields

diflbrInstances :: Maybe Instance

An updated list of remaining instances registered with the load balancer.

data DescribeInstanceHealthResult

The output for the DescribeInstanceHealth action.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_DescribeInstanceHealthResult.html

Constructors

DescribeInstanceHealthResult

Fields

dihrInstanceStates :: Maybe InstanceState

A list containing health information for the specified instances.

data DescribeLoadBalancerPoliciesResult

The output for the DescribeLoadBalancerPolicies action.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_DescribeLoadBalancerPoliciesResult.html

Constructors

DescribeLoadBalancerPoliciesResult

Fields

dlbprPolicyDescriptions :: Maybe PolicyDescription

A list of policy description structures.

data DescribeLoadBalancerPolicyTypesResult

The output for the DescribeLoadBalancerPolicyTypes action.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_DescribeLoadBalancerPolicyTypesResult.html

Constructors

DescribeLoadBalancerPolicyTypesResult

Fields

dlbptrPolicyTypeDescriptions :: Maybe PolicyTypeDescription

List of policy type description structures of the specified policy type. If no policy type names are specified, returns the description of all the policy types defined by Elastic Load Balancing service.

data DescribeLoadBalancersResult

The output for the DescribeLoadBalancers action.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_DescribeLoadBalancersResult.html

Constructors

DescribeLoadBalancersResult

Fields

dlbrLoadBalancerDescriptions :: Maybe LoadBalancerDescription

A list of load balancer description structures.

dlbrNextMarker :: Maybe Text

An optional parameter reserved for future use.

data DetachLoadBalancerFromSubnetsResult

The output for the DetachLoadBalancerFromSubnets action.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_DetachLoadBalancerFromSubnetsResult.html

Constructors

DetachLoadBalancerFromSubnetsResult

Fields

dlbfsrSubnets :: Maybe Text

A list of subnet IDs removed from the configured set of subnets for the load balancer.

data DisableAvailabilityZonesForLoadBalancerResult

The output for the DisableAvailabilityZonesForLoadBalancer action.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_DisableAvailabilityZonesForLoadBalancerResult.html

Constructors

DisableAvailabilityZonesForLoadBalancerResult

Fields

dazflbrAvailabilityZones :: Maybe Text

A list of updated Availability Zones for the load balancer.

data EnableAvailabilityZonesForLoadBalancerResult

The output for the EnableAvailabilityZonesForLoadBalancer action.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_EnableAvailabilityZonesForLoadBalancerResult.html

Constructors

EnableAvailabilityZonesForLoadBalancerResult

Fields

eazflbrAvailabilityZones :: Maybe Text

An updated list of Availability Zones for the load balancer.

data HealthCheck

The HealthCheck data type.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_HealthCheck.html

Constructors

HealthCheck

Fields

hcHealthyThreshold :: !Integer

Specifies the number of consecutive health probe successes required before moving the instance to the Healthy state.

hcInterval :: !Integer

Specifies the approximate interval, in seconds, between health checks of an individual instance.

hcTarget :: !Text

Specifies the instance being checked. The protocol is either TCP, HTTP, HTTPS, or SSL. The range of valid ports is one (1) through 65535.

hcTimeout :: !Integer

Specifies the amount of time, in seconds, during which no response means a failed health probe.

hcUnhealthyThreshold :: !Integer

Specifies the number of consecutive health probe failures required before moving the instance to the Unhealthy state.

data Instance

The Instance data type.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_Instance.html

Constructors

Instance

Fields

iInstanceId :: Maybe Text

Provides an EC2 instance ID.

data InstanceState

The InstanceState data type.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_InstanceState.html

Constructors

InstanceState

Fields

isDescription :: Maybe Text

Provides a description of the instance state.

isInstanceId :: Maybe Text

Provides an EC2 instance ID.

isReasonCode :: Maybe Text

Provides information about the cause of OutOfService instances. Specifically, it indicates whether the cause is Elastic Load Balancing or the instance behind the load balancer.

isState :: Maybe Text

Specifies the current state of the instance. Valid value: InService|OutOfService

data LBCookieStickinessPolicy

The LBCookieStickinessPolicy data type.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_LBCookieStickinessPolicy.html

Constructors

LBCookieStickinessPolicy

Fields

lbcspCookieExpirationPeriod :: Maybe Integer

The time period in seconds after which the cookie should be considered stale. Not specifying this parameter indicates that the stickiness session will last for the duration of the browser session.

lbcspPolicyName :: Maybe Text

The name for the policy being created. The name must be unique within the set of policies for this load balancer.

data Listener

The Listener data type.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_Listener.html

Constructors

Listener

Fields

lInstancePort :: !Integer

Specifies the TCP port on which the instance server is listening. This property cannot be modified for the life of the load balancer.

lInstanceProtocol :: Maybe Text

Specifies the protocol to use for routing traffic to back-end instances - HTTP, HTTPS, TCP, or SSL. This property cannot be modified for the life of the load balancer.

lLoadBalancerPort :: !Integer

Specifies the external load balancer port number. This property cannot be modified for the life of the load balancer.

lProtocol :: !Text

Specifies the load balancer transport protocol to use for routing - HTTP, HTTPS, TCP or SSL. This property cannot be modified for the life of the load balancer.

lSSLCertificateId :: Maybe Text

The ARN string of the server certificate. To get the ARN of the server certificate, call the AWS Identity and Access Management UploadServerCertificate API.

data ListenerDescription

The ListenerDescription data type.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_ListenerDescription.html

Constructors

ListenerDescription

Fields

ldListener :: Maybe Listener

The Listener data type.

ldPolicyNames :: Maybe Text

A list of policies enabled for this listener. An empty list indicates that no policies are enabled.

data LoadBalancerDescription

Contains the result of a successful invocation of DescribeLoadBalancers.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_LoadBalancerDescription.html

Constructors

LoadBalancerDescription

Fields

lbdAvailabilityZones :: Maybe Text

Specifies a list of Availability Zones.

lbdBackendServerDescriptions :: Maybe BackendServerDescription

Contains a list of back-end server descriptions.

lbdCanonicalHostedZoneName :: Maybe Text

Provides the name of the Amazon Route 53 hosted zone that is associated with the load balancer. For information on how to associate your load balancer with a hosted zone, go to Using Domain Names With Elastic Load Balancing in the Elastic Load Balancing Developer Guide.

lbdCanonicalHostedZoneNameID :: Maybe Text

Provides the ID of the Amazon Route 53 hosted zone name that is associated with the load balancer. For information on how to associate or disassociate your load balancer with a hosted zone, go to Using Domain Names With Elastic Load Balancing in the Elastic Load Balancing Developer Guide.

lbdCreatedTime :: Maybe UTCTime

Provides the date and time the load balancer was created.

lbdDNSName :: Maybe Text

Specifies the external DNS name associated with the load balancer.

lbdHealthCheck :: Maybe HealthCheck

Specifies information regarding the various health probes conducted on the load balancer.

lbdInstances :: Maybe Instance

Provides a list of EC2 instance IDs for the load balancer.

lbdListenerDescriptions :: Maybe ListenerDescription

LoadBalancerPort, InstancePort, Protocol, InstanceProtocol, and PolicyNames are returned in a list of tuples in the ListenerDescriptions element.

lbdLoadBalancerName :: Maybe Text

Specifies the name associated with the load balancer.

lbdPolicies :: Maybe Policies

Provides a list of policies defined for the load balancer.

lbdScheme :: Maybe Text

Specifies the type of load balancer.

lbdSecurityGroups :: Maybe Text

The security groups the load balancer is a member of (VPC only).

lbdSourceSecurityGroup :: Maybe SourceSecurityGroup

The security group that you can use as part of your inbound rules for your load balancer's back-end Amazon EC2 application instances. To only allow traffic from load balancers, add a security group rule to your back end instance that specifies this source security group as the inbound source.

lbdSubnets :: Maybe Text

Provides a list of VPC subnet IDs for the load balancer.

lbdVPCId :: Maybe Text

Provides the ID of the VPC attached to the load balancer.

data Policies

The policies data type.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_Policies.html

Constructors

Policies

Fields

pAppCookieStickinessPolicies :: Maybe AppCookieStickinessPolicy

A list of the AppCookieStickinessPolicy objects created with CreateAppCookieStickinessPolicy.

pLBCookieStickinessPolicies :: Maybe LBCookieStickinessPolicy

A list of LBCookieStickinessPolicy objects created with CreateAppCookieStickinessPolicy.

pOtherPolicies :: Maybe Text

A list of policy names other than the stickiness policies.

data PolicyAttribute

The PolicyAttribute data type. This data type contains a key/value pair that defines properties of a specific policy.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_PolicyAttribute.html

Constructors

PolicyAttribute

Fields

paAttributeName :: Maybe Text

The name of the attribute associated with the policy.

paAttributeValue :: Maybe Text

The value of the attribute associated with the policy.

data PolicyAttributeDescription

The PolicyAttributeDescription data type. This data type is used to describe the attributes and values associated with a policy.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_PolicyAttributeDescription.html

Constructors

PolicyAttributeDescription

Fields

padAttributeName :: Maybe Text

The name of the attribute associated with the policy.

padAttributeValue :: Maybe Text

The value of the attribute associated with the policy.

data PolicyAttributeTypeDescription

The PolicyAttributeTypeDescription data type. This data type is used to describe values that are acceptable for the policy attribute.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_PolicyAttributeTypeDescription.html

Constructors

PolicyAttributeTypeDescription

Fields

patdAttributeName :: Maybe Text

The name of the attribute associated with the policy type.

patdAttributeType :: Maybe Text

The type of attribute. For example, Boolean, Integer, etc.

patdCardinality :: Maybe Text

The cardinality of the attribute. Valid Values:

patdDefaultValue :: Maybe Text

The default value of the attribute, if applicable.

patdDescription :: Maybe Text

A human-readable description of the attribute.

data PolicyDescription

The PolicyDescription data type.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_PolicyDescription.html

Constructors

PolicyDescription

Fields

pdPolicyAttributeDescriptions :: Maybe PolicyAttributeDescription

A list of policy attribute description structures.

pdPolicyName :: Maybe Text

The name of the policy associated with the load balancer.

pdPolicyTypeName :: Maybe Text

The name of the policy type associated with the load balancer.

data PolicyTypeDescription

The PolicyTypeDescription data type.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_PolicyTypeDescription.html

Constructors

PolicyTypeDescription

Fields

ptdDescription :: Maybe Text

A human-readable description of the policy type.

ptdPolicyAttributeTypeDescriptions :: Maybe PolicyAttributeTypeDescription

The description of the policy attributes associated with the load balancer policies defined by the Elastic Load Balancing service.

ptdPolicyTypeName :: Maybe Text

The name of the policy type.

data RegisterInstancesWithLoadBalancerResult

The output for the RegisterInstancesWithLoadBalancer action.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_RegisterInstancesWithLoadBalancerResult.html

Constructors

RegisterInstancesWithLoadBalancerResult

Fields

riwlbrInstances :: Maybe Instance

An updated list of instances for the load balancer.

data SetLoadBalancerListenerSSLCertificateResult

The output for the SetLoadBalancerListenerSSLCertificate action.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_SetLoadBalancerListenerSSLCertificateResult.html

Constructors

SetLoadBalancerListenerSSLCertificateResult

data SetLoadBalancerPoliciesForBackendServerResult

The output for the SetLoadBalancerPoliciesForBackendServer action.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_SetLoadBalancerPoliciesForBackendServerResult.html

Constructors

SetLoadBalancerPoliciesForBackendServerResult

data SetLoadBalancerPoliciesOfListenerResult

The output for the SetLoadBalancerPoliciesOfListener action.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_SetLoadBalancerPoliciesOfListenerResult.html

Constructors

SetLoadBalancerPoliciesOfListenerResult

data SourceSecurityGroup

This data type is used as a response element in the DescribeLoadBalancers action. For information about Elastic Load Balancing security groups, go to Using Security Groups With Elastic Load Balancing in the Elastic Load Balancing Developer Guide.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_SourceSecurityGroup.html

Constructors

SourceSecurityGroup

Fields

ssgGroupName :: Maybe Text

Name of the source security group. Use this value for the --source-group parameter of the ec2-authorize command in the Amazon EC2 command line tool.

ssgOwnerAlias :: Maybe Text

Owner of the source security group. Use this value for the --source-group-user parameter of the ec2-authorize command in the Amazon EC2 command line tool.

elbNS :: ByteString

XML namespace to annotate ELB elements with.

elbElem :: ByteString -> NName ByteString

Helper to define ELB namespaced XML elements.

Common

module Network.AWS