Network.AWS.ELB.AttachLoadBalancerToSubnets

Description

Adds one or more subnets to the set of configured subnets for the specified load balancer.

The load balancer evenly distributes requests across all registered subnets. For more information, see Add or Remove Subnets for Your Load Balancer in a VPC in the Classic Load Balancers Guide.

Creating a Request

data AttachLoadBalancerToSubnets

Contains the parameters for AttachLoaBalancerToSubnets.

See: newAttachLoadBalancerToSubnets smart constructor.

Constructors

AttachLoadBalancerToSubnets'

Fields

loadBalancerName :: Text

The name of the load balancer.

subnets :: [Text]

The IDs of the subnets to add. You can add only one subnet per Availability Zone.

newAttachLoadBalancerToSubnets

Arguments

:: Text
$sel:loadBalancerName:AttachLoadBalancerToSubnets'

-> AttachLoadBalancerToSubnets

Create a value of AttachLoadBalancerToSubnets with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:loadBalancerName:AttachLoadBalancerToSubnets', attachLoadBalancerToSubnets_loadBalancerName - The name of the load balancer.

$sel:subnets:AttachLoadBalancerToSubnets', attachLoadBalancerToSubnets_subnets - The IDs of the subnets to add. You can add only one subnet per Availability Zone.

Request Lenses

attachLoadBalancerToSubnets_loadBalancerName :: Lens' AttachLoadBalancerToSubnets Text

The name of the load balancer.

attachLoadBalancerToSubnets_subnets :: Lens' AttachLoadBalancerToSubnets [Text]

The IDs of the subnets to add. You can add only one subnet per Availability Zone.

Destructuring the Response

data AttachLoadBalancerToSubnetsResponse

Contains the output of AttachLoadBalancerToSubnets.

See: newAttachLoadBalancerToSubnetsResponse smart constructor.

Constructors

AttachLoadBalancerToSubnetsResponse'

Fields

subnets :: Maybe [Text]

The IDs of the subnets attached to the load balancer.

httpStatus :: Int

The response's http status code.

newAttachLoadBalancerToSubnetsResponse

Arguments

:: Int
$sel:httpStatus:AttachLoadBalancerToSubnetsResponse'

-> AttachLoadBalancerToSubnetsResponse

Create a value of AttachLoadBalancerToSubnetsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:subnets:AttachLoadBalancerToSubnets', attachLoadBalancerToSubnetsResponse_subnets - The IDs of the subnets attached to the load balancer.

$sel:httpStatus:AttachLoadBalancerToSubnetsResponse', attachLoadBalancerToSubnetsResponse_httpStatus - The response's http status code.

Response Lenses

attachLoadBalancerToSubnetsResponse_subnets :: Lens' AttachLoadBalancerToSubnetsResponse (Maybe [Text])

The IDs of the subnets attached to the load balancer.

attachLoadBalancerToSubnetsResponse_httpStatus :: Lens' AttachLoadBalancerToSubnetsResponse Int

The response's http status code.