Network.AWS.ELB.CreateLoadBalancer

Description

Creates a Classic Load Balancer.

You can add listeners, security groups, subnets, and tags when you create your load balancer, or you can add them later using CreateLoadBalancerListeners, ApplySecurityGroupsToLoadBalancer, AttachLoadBalancerToSubnets, and AddTags.

To describe your current load balancers, see DescribeLoadBalancers. When you are finished with a load balancer, you can delete it using DeleteLoadBalancer.

You can create up to 20 load balancers per region per account. You can request an increase for the number of load balancers for your account. For more information, see Limits for Your Classic Load Balancer in the Classic Load Balancers Guide.

Creating a Request

data CreateLoadBalancer

Contains the parameters for CreateLoadBalancer.

See: newCreateLoadBalancer smart constructor.

Constructors

CreateLoadBalancer'

Fields

availabilityZones :: Maybe [Text]

One or more Availability Zones from the same region as the load balancer.

You must specify at least one Availability Zone.

You can add more Availability Zones after you create the load balancer using EnableAvailabilityZonesForLoadBalancer.

scheme :: Maybe Text

The type of a load balancer. Valid only for load balancers in a VPC.

By default, Elastic Load Balancing creates an Internet-facing load balancer with a DNS name that resolves to public IP addresses. For more information about Internet-facing and Internal load balancers, see Load Balancer Scheme in the Elastic Load Balancing User Guide.

Specify internal to create a load balancer with a DNS name that resolves to private IP addresses.

securityGroups :: Maybe [Text]

The IDs of the security groups to assign to the load balancer.

tags :: Maybe (NonEmpty Tag)

A list of tags to assign to the load balancer.

For more information about tagging your load balancer, see Tag Your Classic Load Balancer in the Classic Load Balancers Guide.

subnets :: Maybe [Text]

The IDs of the subnets in your VPC to attach to the load balancer. Specify one subnet per Availability Zone specified in AvailabilityZones.

loadBalancerName :: Text

The name of the load balancer.

This name must be unique within your set of load balancers for the region, must have a maximum of 32 characters, must contain only alphanumeric characters or hyphens, and cannot begin or end with a hyphen.

listeners :: [Listener]

The listeners.

For more information, see Listeners for Your Classic Load Balancer in the Classic Load Balancers Guide.

newCreateLoadBalancer

Arguments

:: Text
$sel:loadBalancerName:CreateLoadBalancer'

-> CreateLoadBalancer

Create a value of CreateLoadBalancer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:availabilityZones:CreateLoadBalancer', createLoadBalancer_availabilityZones - One or more Availability Zones from the same region as the load balancer.

You must specify at least one Availability Zone.

You can add more Availability Zones after you create the load balancer using EnableAvailabilityZonesForLoadBalancer.

$sel:scheme:CreateLoadBalancer', createLoadBalancer_scheme - The type of a load balancer. Valid only for load balancers in a VPC.

By default, Elastic Load Balancing creates an Internet-facing load balancer with a DNS name that resolves to public IP addresses. For more information about Internet-facing and Internal load balancers, see Load Balancer Scheme in the Elastic Load Balancing User Guide.

Specify internal to create a load balancer with a DNS name that resolves to private IP addresses.

$sel:securityGroups:CreateLoadBalancer', createLoadBalancer_securityGroups - The IDs of the security groups to assign to the load balancer.

$sel:tags:CreateLoadBalancer', createLoadBalancer_tags - A list of tags to assign to the load balancer.

For more information about tagging your load balancer, see Tag Your Classic Load Balancer in the Classic Load Balancers Guide.

$sel:subnets:CreateLoadBalancer', createLoadBalancer_subnets - The IDs of the subnets in your VPC to attach to the load balancer. Specify one subnet per Availability Zone specified in AvailabilityZones.

$sel:loadBalancerName:CreateLoadBalancer', createLoadBalancer_loadBalancerName - The name of the load balancer.

This name must be unique within your set of load balancers for the region, must have a maximum of 32 characters, must contain only alphanumeric characters or hyphens, and cannot begin or end with a hyphen.

$sel:listeners:CreateLoadBalancer', createLoadBalancer_listeners - The listeners.

For more information, see Listeners for Your Classic Load Balancer in the Classic Load Balancers Guide.

Request Lenses

createLoadBalancer_availabilityZones :: Lens' CreateLoadBalancer (Maybe [Text])

One or more Availability Zones from the same region as the load balancer.

You must specify at least one Availability Zone.

You can add more Availability Zones after you create the load balancer using EnableAvailabilityZonesForLoadBalancer.

createLoadBalancer_scheme :: Lens' CreateLoadBalancer (Maybe Text)

The type of a load balancer. Valid only for load balancers in a VPC.

By default, Elastic Load Balancing creates an Internet-facing load balancer with a DNS name that resolves to public IP addresses. For more information about Internet-facing and Internal load balancers, see Load Balancer Scheme in the Elastic Load Balancing User Guide.

Specify internal to create a load balancer with a DNS name that resolves to private IP addresses.

createLoadBalancer_securityGroups :: Lens' CreateLoadBalancer (Maybe [Text])

The IDs of the security groups to assign to the load balancer.

createLoadBalancer_tags :: Lens' CreateLoadBalancer (Maybe (NonEmpty Tag))

A list of tags to assign to the load balancer.

For more information about tagging your load balancer, see Tag Your Classic Load Balancer in the Classic Load Balancers Guide.

createLoadBalancer_subnets :: Lens' CreateLoadBalancer (Maybe [Text])

The IDs of the subnets in your VPC to attach to the load balancer. Specify one subnet per Availability Zone specified in AvailabilityZones.

createLoadBalancer_loadBalancerName :: Lens' CreateLoadBalancer Text

The name of the load balancer.

This name must be unique within your set of load balancers for the region, must have a maximum of 32 characters, must contain only alphanumeric characters or hyphens, and cannot begin or end with a hyphen.

createLoadBalancer_listeners :: Lens' CreateLoadBalancer [Listener]

The listeners.

For more information, see Listeners for Your Classic Load Balancer in the Classic Load Balancers Guide.

Destructuring the Response

data CreateLoadBalancerResponse

Contains the output for CreateLoadBalancer.

See: newCreateLoadBalancerResponse smart constructor.

Constructors

CreateLoadBalancerResponse'

Fields

dNSName :: Maybe Text

The DNS name of the load balancer.

httpStatus :: Int

The response's http status code.

newCreateLoadBalancerResponse

Arguments

:: Int
$sel:httpStatus:CreateLoadBalancerResponse'

-> CreateLoadBalancerResponse

Create a value of CreateLoadBalancerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dNSName:CreateLoadBalancerResponse', createLoadBalancerResponse_dNSName - The DNS name of the load balancer.

$sel:httpStatus:CreateLoadBalancerResponse', createLoadBalancerResponse_httpStatus - The response's http status code.

Response Lenses

createLoadBalancerResponse_dNSName :: Lens' CreateLoadBalancerResponse (Maybe Text)

The DNS name of the load balancer.

createLoadBalancerResponse_httpStatus :: Lens' CreateLoadBalancerResponse Int

The response's http status code.