Network.AWS.ELB.CreateLoadBalancerListeners

Description

Creates one or more listeners for the specified load balancer. If a listener with the specified port does not already exist, it is created; otherwise, the properties of the new listener must match the properties of the existing listener.

For more information, see Listeners for Your Classic Load Balancer in the Classic Load Balancers Guide.

Creating a Request

data CreateLoadBalancerListeners

Contains the parameters for CreateLoadBalancerListeners.

See: newCreateLoadBalancerListeners smart constructor.

Constructors

CreateLoadBalancerListeners'

Fields

loadBalancerName :: Text

The name of the load balancer.

listeners :: [Listener]

The listeners.

newCreateLoadBalancerListeners

Arguments

:: Text
$sel:loadBalancerName:CreateLoadBalancerListeners'

-> CreateLoadBalancerListeners

Create a value of CreateLoadBalancerListeners with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:loadBalancerName:CreateLoadBalancerListeners', createLoadBalancerListeners_loadBalancerName - The name of the load balancer.

$sel:listeners:CreateLoadBalancerListeners', createLoadBalancerListeners_listeners - The listeners.

Request Lenses

createLoadBalancerListeners_loadBalancerName :: Lens' CreateLoadBalancerListeners Text

The name of the load balancer.

createLoadBalancerListeners_listeners :: Lens' CreateLoadBalancerListeners [Listener]

The listeners.

Destructuring the Response

data CreateLoadBalancerListenersResponse

Contains the parameters for CreateLoadBalancerListener.

See: newCreateLoadBalancerListenersResponse smart constructor.

Constructors

CreateLoadBalancerListenersResponse'

Fields

httpStatus :: Int

The response's http status code.

newCreateLoadBalancerListenersResponse

Arguments

:: Int
$sel:httpStatus:CreateLoadBalancerListenersResponse'

-> CreateLoadBalancerListenersResponse

Create a value of CreateLoadBalancerListenersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateLoadBalancerListenersResponse', createLoadBalancerListenersResponse_httpStatus - The response's http status code.

Response Lenses

createLoadBalancerListenersResponse_httpStatus :: Lens' CreateLoadBalancerListenersResponse Int

The response's http status code.