Network.AWS.ELB.CreateLoadBalancerPolicy

Description

Creates a policy with the specified attributes for the specified load balancer.

Policies are settings that are saved for your load balancer and that can be applied to the listener or the application server, depending on the policy type.

Creating a Request

data CreateLoadBalancerPolicy

Contains the parameters for CreateLoadBalancerPolicy.

See: newCreateLoadBalancerPolicy smart constructor.

Constructors

CreateLoadBalancerPolicy'

Fields

policyAttributes :: Maybe [PolicyAttribute]

The policy attributes.

loadBalancerName :: Text

The name of the load balancer.

policyName :: Text

The name of the load balancer policy to be created. This name must be unique within the set of policies for this load balancer.

policyTypeName :: Text

The name of the base policy type. To get the list of policy types, use DescribeLoadBalancerPolicyTypes.

newCreateLoadBalancerPolicy

Arguments

:: Text
$sel:loadBalancerName:CreateLoadBalancerPolicy'

-> Text
$sel:policyName:CreateLoadBalancerPolicy'

-> Text
$sel:policyTypeName:CreateLoadBalancerPolicy'

-> CreateLoadBalancerPolicy

Create a value of CreateLoadBalancerPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policyAttributes:CreateLoadBalancerPolicy', createLoadBalancerPolicy_policyAttributes - The policy attributes.

$sel:loadBalancerName:CreateLoadBalancerPolicy', createLoadBalancerPolicy_loadBalancerName - The name of the load balancer.

$sel:policyName:CreateLoadBalancerPolicy', createLoadBalancerPolicy_policyName - The name of the load balancer policy to be created. This name must be unique within the set of policies for this load balancer.

$sel:policyTypeName:CreateLoadBalancerPolicy', createLoadBalancerPolicy_policyTypeName - The name of the base policy type. To get the list of policy types, use DescribeLoadBalancerPolicyTypes.

Request Lenses

createLoadBalancerPolicy_policyAttributes :: Lens' CreateLoadBalancerPolicy (Maybe [PolicyAttribute])

The policy attributes.

createLoadBalancerPolicy_loadBalancerName :: Lens' CreateLoadBalancerPolicy Text

The name of the load balancer.

createLoadBalancerPolicy_policyName :: Lens' CreateLoadBalancerPolicy Text

The name of the load balancer policy to be created. This name must be unique within the set of policies for this load balancer.

createLoadBalancerPolicy_policyTypeName :: Lens' CreateLoadBalancerPolicy Text

The name of the base policy type. To get the list of policy types, use DescribeLoadBalancerPolicyTypes.

Destructuring the Response

data CreateLoadBalancerPolicyResponse

Contains the output of CreateLoadBalancerPolicy.

See: newCreateLoadBalancerPolicyResponse smart constructor.

Constructors

CreateLoadBalancerPolicyResponse'

Fields

httpStatus :: Int

The response's http status code.

newCreateLoadBalancerPolicyResponse

Arguments

:: Int
$sel:httpStatus:CreateLoadBalancerPolicyResponse'

-> CreateLoadBalancerPolicyResponse

Create a value of CreateLoadBalancerPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateLoadBalancerPolicyResponse', createLoadBalancerPolicyResponse_httpStatus - The response's http status code.

Response Lenses

createLoadBalancerPolicyResponse_httpStatus :: Lens' CreateLoadBalancerPolicyResponse Int

The response's http status code.