Network.AWS.ELB.DeleteLoadBalancerListeners

Description

Deletes the specified listeners from the specified load balancer.

Creating a Request

data DeleteLoadBalancerListeners

Contains the parameters for DeleteLoadBalancerListeners.

See: newDeleteLoadBalancerListeners smart constructor.

Constructors

DeleteLoadBalancerListeners'

Fields

loadBalancerName :: Text

The name of the load balancer.

loadBalancerPorts :: [Int]

The client port numbers of the listeners.

newDeleteLoadBalancerListeners

Arguments

:: Text
$sel:loadBalancerName:DeleteLoadBalancerListeners'

-> DeleteLoadBalancerListeners

Create a value of DeleteLoadBalancerListeners with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:loadBalancerName:DeleteLoadBalancerListeners', deleteLoadBalancerListeners_loadBalancerName - The name of the load balancer.

$sel:loadBalancerPorts:DeleteLoadBalancerListeners', deleteLoadBalancerListeners_loadBalancerPorts - The client port numbers of the listeners.

Request Lenses

deleteLoadBalancerListeners_loadBalancerName :: Lens' DeleteLoadBalancerListeners Text

The name of the load balancer.

deleteLoadBalancerListeners_loadBalancerPorts :: Lens' DeleteLoadBalancerListeners [Int]

The client port numbers of the listeners.

Destructuring the Response

data DeleteLoadBalancerListenersResponse

Contains the output of DeleteLoadBalancerListeners.

See: newDeleteLoadBalancerListenersResponse smart constructor.

Constructors

DeleteLoadBalancerListenersResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteLoadBalancerListenersResponse

Arguments

:: Int
$sel:httpStatus:DeleteLoadBalancerListenersResponse'

-> DeleteLoadBalancerListenersResponse

Create a value of DeleteLoadBalancerListenersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteLoadBalancerListenersResponse', deleteLoadBalancerListenersResponse_httpStatus - The response's http status code.

Response Lenses

deleteLoadBalancerListenersResponse_httpStatus :: Lens' DeleteLoadBalancerListenersResponse Int

The response's http status code.