Network.AWS.ELB.DescribeInstanceHealth

Description

Describes the state of the specified instances with respect to the specified load balancer. If no instances are specified, the call describes the state of all instances that are currently registered with the load balancer. If instances are specified, their state is returned even if they are no longer registered with the load balancer. The state of terminated instances is not returned.

Creating a Request

data DescribeInstanceHealth

Contains the parameters for DescribeInstanceHealth.

See: newDescribeInstanceHealth smart constructor.

Constructors

DescribeInstanceHealth'

Fields

instances :: Maybe [Instance]

The IDs of the instances.

loadBalancerName :: Text

The name of the load balancer.

newDescribeInstanceHealth

Arguments

:: Text
$sel:loadBalancerName:DescribeInstanceHealth'

-> DescribeInstanceHealth

Create a value of DescribeInstanceHealth with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instances:DescribeInstanceHealth', describeInstanceHealth_instances - The IDs of the instances.

$sel:loadBalancerName:DescribeInstanceHealth', describeInstanceHealth_loadBalancerName - The name of the load balancer.

Request Lenses

describeInstanceHealth_instances :: Lens' DescribeInstanceHealth (Maybe [Instance])

The IDs of the instances.

describeInstanceHealth_loadBalancerName :: Lens' DescribeInstanceHealth Text

The name of the load balancer.

Destructuring the Response

data DescribeInstanceHealthResponse

Contains the output for DescribeInstanceHealth.

See: newDescribeInstanceHealthResponse smart constructor.

Constructors

DescribeInstanceHealthResponse'

Fields

instanceStates :: Maybe [InstanceState]

Information about the health of the instances.

httpStatus :: Int

The response's http status code.

newDescribeInstanceHealthResponse

Arguments

:: Int
$sel:httpStatus:DescribeInstanceHealthResponse'

-> DescribeInstanceHealthResponse

Create a value of DescribeInstanceHealthResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceStates:DescribeInstanceHealthResponse', describeInstanceHealthResponse_instanceStates - Information about the health of the instances.

$sel:httpStatus:DescribeInstanceHealthResponse', describeInstanceHealthResponse_httpStatus - The response's http status code.

Response Lenses

describeInstanceHealthResponse_instanceStates :: Lens' DescribeInstanceHealthResponse (Maybe [InstanceState])

Information about the health of the instances.

describeInstanceHealthResponse_httpStatus :: Lens' DescribeInstanceHealthResponse Int

The response's http status code.