Network.AWS.ELB.DisableAvailabilityZonesForLoadBalancer

Description

Removes the specified Availability Zones from the set of Availability Zones for the specified load balancer in EC2-Classic or a default VPC.

For load balancers in a non-default VPC, use DetachLoadBalancerFromSubnets.

There must be at least one Availability Zone registered with a load balancer at all times. After an Availability Zone is removed, all instances registered with the load balancer that are in the removed Availability Zone go into the OutOfService state. Then, the load balancer attempts to equally balance the traffic among its remaining Availability Zones.

For more information, see Add or Remove Availability Zones in the Classic Load Balancers Guide.

Creating a Request

data DisableAvailabilityZonesForLoadBalancer

Contains the parameters for DisableAvailabilityZonesForLoadBalancer.

See: newDisableAvailabilityZonesForLoadBalancer smart constructor.

Constructors

DisableAvailabilityZonesForLoadBalancer'

Fields

loadBalancerName :: Text

The name of the load balancer.

availabilityZones :: [Text]

The Availability Zones.

newDisableAvailabilityZonesForLoadBalancer

Arguments

:: Text
$sel:loadBalancerName:DisableAvailabilityZonesForLoadBalancer'

-> DisableAvailabilityZonesForLoadBalancer

Create a value of DisableAvailabilityZonesForLoadBalancer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:loadBalancerName:DisableAvailabilityZonesForLoadBalancer', disableAvailabilityZonesForLoadBalancer_loadBalancerName - The name of the load balancer.

$sel:availabilityZones:DisableAvailabilityZonesForLoadBalancer', disableAvailabilityZonesForLoadBalancer_availabilityZones - The Availability Zones.

Request Lenses

disableAvailabilityZonesForLoadBalancer_loadBalancerName :: Lens' DisableAvailabilityZonesForLoadBalancer Text

The name of the load balancer.

disableAvailabilityZonesForLoadBalancer_availabilityZones :: Lens' DisableAvailabilityZonesForLoadBalancer [Text]

The Availability Zones.

Destructuring the Response

data DisableAvailabilityZonesForLoadBalancerResponse

Contains the output for DisableAvailabilityZonesForLoadBalancer.

See: newDisableAvailabilityZonesForLoadBalancerResponse smart constructor.

Constructors

DisableAvailabilityZonesForLoadBalancerResponse'

Fields

availabilityZones :: Maybe [Text]

The remaining Availability Zones for the load balancer.

httpStatus :: Int

The response's http status code.

newDisableAvailabilityZonesForLoadBalancerResponse

Arguments

:: Int
$sel:httpStatus:DisableAvailabilityZonesForLoadBalancerResponse'

-> DisableAvailabilityZonesForLoadBalancerResponse

Create a value of DisableAvailabilityZonesForLoadBalancerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:availabilityZones:DisableAvailabilityZonesForLoadBalancer', disableAvailabilityZonesForLoadBalancerResponse_availabilityZones - The remaining Availability Zones for the load balancer.

$sel:httpStatus:DisableAvailabilityZonesForLoadBalancerResponse', disableAvailabilityZonesForLoadBalancerResponse_httpStatus - The response's http status code.

Response Lenses

disableAvailabilityZonesForLoadBalancerResponse_availabilityZones :: Lens' DisableAvailabilityZonesForLoadBalancerResponse (Maybe [Text])

The remaining Availability Zones for the load balancer.

disableAvailabilityZonesForLoadBalancerResponse_httpStatus :: Lens' DisableAvailabilityZonesForLoadBalancerResponse Int

The response's http status code.