Network.AWS.ELB.EnableAvailabilityZonesForLoadBalancer

Description

Adds the specified Availability Zones to the set of Availability Zones for the specified load balancer in EC2-Classic or a default VPC.

For load balancers in a non-default VPC, use AttachLoadBalancerToSubnets.

The load balancer evenly distributes requests across all its registered Availability Zones that contain instances. For more information, see Add or Remove Availability Zones in the Classic Load Balancers Guide.

Creating a Request

data EnableAvailabilityZonesForLoadBalancer

Contains the parameters for EnableAvailabilityZonesForLoadBalancer.

See: newEnableAvailabilityZonesForLoadBalancer smart constructor.

Constructors

EnableAvailabilityZonesForLoadBalancer'

Fields

loadBalancerName :: Text

The name of the load balancer.

availabilityZones :: [Text]

The Availability Zones. These must be in the same region as the load balancer.

newEnableAvailabilityZonesForLoadBalancer

Arguments

:: Text
$sel:loadBalancerName:EnableAvailabilityZonesForLoadBalancer'

-> EnableAvailabilityZonesForLoadBalancer

Create a value of EnableAvailabilityZonesForLoadBalancer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:loadBalancerName:EnableAvailabilityZonesForLoadBalancer', enableAvailabilityZonesForLoadBalancer_loadBalancerName - The name of the load balancer.

$sel:availabilityZones:EnableAvailabilityZonesForLoadBalancer', enableAvailabilityZonesForLoadBalancer_availabilityZones - The Availability Zones. These must be in the same region as the load balancer.

Request Lenses

enableAvailabilityZonesForLoadBalancer_loadBalancerName :: Lens' EnableAvailabilityZonesForLoadBalancer Text

The name of the load balancer.

enableAvailabilityZonesForLoadBalancer_availabilityZones :: Lens' EnableAvailabilityZonesForLoadBalancer [Text]

The Availability Zones. These must be in the same region as the load balancer.

Destructuring the Response

data EnableAvailabilityZonesForLoadBalancerResponse

Contains the output of EnableAvailabilityZonesForLoadBalancer.

See: newEnableAvailabilityZonesForLoadBalancerResponse smart constructor.

Constructors

EnableAvailabilityZonesForLoadBalancerResponse'

Fields

availabilityZones :: Maybe [Text]

The updated list of Availability Zones for the load balancer.

httpStatus :: Int

The response's http status code.

newEnableAvailabilityZonesForLoadBalancerResponse

Arguments

:: Int
$sel:httpStatus:EnableAvailabilityZonesForLoadBalancerResponse'

-> EnableAvailabilityZonesForLoadBalancerResponse

Create a value of EnableAvailabilityZonesForLoadBalancerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:availabilityZones:EnableAvailabilityZonesForLoadBalancer', enableAvailabilityZonesForLoadBalancerResponse_availabilityZones - The updated list of Availability Zones for the load balancer.

$sel:httpStatus:EnableAvailabilityZonesForLoadBalancerResponse', enableAvailabilityZonesForLoadBalancerResponse_httpStatus - The response's http status code.

Response Lenses

enableAvailabilityZonesForLoadBalancerResponse_availabilityZones :: Lens' EnableAvailabilityZonesForLoadBalancerResponse (Maybe [Text])

The updated list of Availability Zones for the load balancer.

enableAvailabilityZonesForLoadBalancerResponse_httpStatus :: Lens' EnableAvailabilityZonesForLoadBalancerResponse Int

The response's http status code.