Network.AWS.ELB.Lens

Description

Operations

SetLoadBalancerListenerSSLCertificate

setLoadBalancerListenerSSLCertificate_loadBalancerName :: Lens' SetLoadBalancerListenerSSLCertificate Text

The name of the load balancer.

setLoadBalancerListenerSSLCertificate_loadBalancerPort :: Lens' SetLoadBalancerListenerSSLCertificate Int

The port that uses the specified SSL certificate.

setLoadBalancerListenerSSLCertificate_sSLCertificateId :: Lens' SetLoadBalancerListenerSSLCertificate Text

The Amazon Resource Name (ARN) of the SSL certificate.

setLoadBalancerListenerSSLCertificateResponse_httpStatus :: Lens' SetLoadBalancerListenerSSLCertificateResponse Int

The response's http status code.

RemoveTags

removeTags_loadBalancerNames :: Lens' RemoveTags [Text]

The name of the load balancer. You can specify a maximum of one load balancer name.

removeTags_tags :: Lens' RemoveTags (NonEmpty TagKeyOnly)

The list of tag keys to remove.

removeTagsResponse_httpStatus :: Lens' RemoveTagsResponse Int

The response's http status code.

DescribeTags

describeTags_loadBalancerNames :: Lens' DescribeTags (NonEmpty Text)

The names of the load balancers.

describeTagsResponse_tagDescriptions :: Lens' DescribeTagsResponse (Maybe [TagDescription])

Information about the tags.

describeTagsResponse_httpStatus :: Lens' DescribeTagsResponse Int

The response's http status code.

SetLoadBalancerPoliciesForBackendServer

setLoadBalancerPoliciesForBackendServer_loadBalancerName :: Lens' SetLoadBalancerPoliciesForBackendServer Text

The name of the load balancer.

setLoadBalancerPoliciesForBackendServer_instancePort :: Lens' SetLoadBalancerPoliciesForBackendServer Int

The port number associated with the EC2 instance.

setLoadBalancerPoliciesForBackendServer_policyNames :: Lens' SetLoadBalancerPoliciesForBackendServer [Text]

The names of the policies. If the list is empty, then all current polices are removed from the EC2 instance.

setLoadBalancerPoliciesForBackendServerResponse_httpStatus :: Lens' SetLoadBalancerPoliciesForBackendServerResponse Int

The response's http status code.

SetLoadBalancerPoliciesOfListener

setLoadBalancerPoliciesOfListener_loadBalancerName :: Lens' SetLoadBalancerPoliciesOfListener Text

The name of the load balancer.

setLoadBalancerPoliciesOfListener_loadBalancerPort :: Lens' SetLoadBalancerPoliciesOfListener Int

The external port of the load balancer.

setLoadBalancerPoliciesOfListener_policyNames :: Lens' SetLoadBalancerPoliciesOfListener [Text]

The names of the policies. This list must include all policies to be enabled. If you omit a policy that is currently enabled, it is disabled. If the list is empty, all current policies are disabled.

setLoadBalancerPoliciesOfListenerResponse_httpStatus :: Lens' SetLoadBalancerPoliciesOfListenerResponse Int

The response's http status code.

DescribeLoadBalancerPolicies

describeLoadBalancerPolicies_policyNames :: Lens' DescribeLoadBalancerPolicies (Maybe [Text])

The names of the policies.

describeLoadBalancerPolicies_loadBalancerName :: Lens' DescribeLoadBalancerPolicies (Maybe Text)

The name of the load balancer.

describeLoadBalancerPoliciesResponse_policyDescriptions :: Lens' DescribeLoadBalancerPoliciesResponse (Maybe [PolicyDescription])

Information about the policies.

describeLoadBalancerPoliciesResponse_httpStatus :: Lens' DescribeLoadBalancerPoliciesResponse Int

The response's http status code.

CreateLoadBalancer

createLoadBalancer_availabilityZones :: Lens' CreateLoadBalancer (Maybe [Text])

One or more Availability Zones from the same region as the load balancer.

You must specify at least one Availability Zone.

You can add more Availability Zones after you create the load balancer using EnableAvailabilityZonesForLoadBalancer.

createLoadBalancer_scheme :: Lens' CreateLoadBalancer (Maybe Text)

The type of a load balancer. Valid only for load balancers in a VPC.

By default, Elastic Load Balancing creates an Internet-facing load balancer with a DNS name that resolves to public IP addresses. For more information about Internet-facing and Internal load balancers, see Load Balancer Scheme in the Elastic Load Balancing User Guide.

Specify internal to create a load balancer with a DNS name that resolves to private IP addresses.

createLoadBalancer_securityGroups :: Lens' CreateLoadBalancer (Maybe [Text])

The IDs of the security groups to assign to the load balancer.

createLoadBalancer_tags :: Lens' CreateLoadBalancer (Maybe (NonEmpty Tag))

A list of tags to assign to the load balancer.

For more information about tagging your load balancer, see Tag Your Classic Load Balancer in the Classic Load Balancers Guide.

createLoadBalancer_subnets :: Lens' CreateLoadBalancer (Maybe [Text])

The IDs of the subnets in your VPC to attach to the load balancer. Specify one subnet per Availability Zone specified in AvailabilityZones.

createLoadBalancer_loadBalancerName :: Lens' CreateLoadBalancer Text

The name of the load balancer.

This name must be unique within your set of load balancers for the region, must have a maximum of 32 characters, must contain only alphanumeric characters or hyphens, and cannot begin or end with a hyphen.

createLoadBalancer_listeners :: Lens' CreateLoadBalancer [Listener]

The listeners.

For more information, see Listeners for Your Classic Load Balancer in the Classic Load Balancers Guide.

createLoadBalancerResponse_dNSName :: Lens' CreateLoadBalancerResponse (Maybe Text)

The DNS name of the load balancer.

createLoadBalancerResponse_httpStatus :: Lens' CreateLoadBalancerResponse Int

The response's http status code.

DeregisterInstancesFromLoadBalancer

deregisterInstancesFromLoadBalancer_loadBalancerName :: Lens' DeregisterInstancesFromLoadBalancer Text

The name of the load balancer.

deregisterInstancesFromLoadBalancer_instances :: Lens' DeregisterInstancesFromLoadBalancer [Instance]

The IDs of the instances.

deregisterInstancesFromLoadBalancerResponse_instances :: Lens' DeregisterInstancesFromLoadBalancerResponse (Maybe [Instance])

The remaining instances registered with the load balancer.

deregisterInstancesFromLoadBalancerResponse_httpStatus :: Lens' DeregisterInstancesFromLoadBalancerResponse Int

The response's http status code.

RegisterInstancesWithLoadBalancer

registerInstancesWithLoadBalancer_loadBalancerName :: Lens' RegisterInstancesWithLoadBalancer Text

The name of the load balancer.

registerInstancesWithLoadBalancer_instances :: Lens' RegisterInstancesWithLoadBalancer [Instance]

The IDs of the instances.

registerInstancesWithLoadBalancerResponse_instances :: Lens' RegisterInstancesWithLoadBalancerResponse (Maybe [Instance])

The updated list of instances for the load balancer.

registerInstancesWithLoadBalancerResponse_httpStatus :: Lens' RegisterInstancesWithLoadBalancerResponse Int

The response's http status code.

DetachLoadBalancerFromSubnets

detachLoadBalancerFromSubnets_loadBalancerName :: Lens' DetachLoadBalancerFromSubnets Text

The name of the load balancer.

detachLoadBalancerFromSubnets_subnets :: Lens' DetachLoadBalancerFromSubnets [Text]

The IDs of the subnets.

detachLoadBalancerFromSubnetsResponse_subnets :: Lens' DetachLoadBalancerFromSubnetsResponse (Maybe [Text])

The IDs of the remaining subnets for the load balancer.

detachLoadBalancerFromSubnetsResponse_httpStatus :: Lens' DetachLoadBalancerFromSubnetsResponse Int

The response's http status code.

CreateLoadBalancerListeners

createLoadBalancerListeners_loadBalancerName :: Lens' CreateLoadBalancerListeners Text

The name of the load balancer.

createLoadBalancerListeners_listeners :: Lens' CreateLoadBalancerListeners [Listener]

The listeners.

createLoadBalancerListenersResponse_httpStatus :: Lens' CreateLoadBalancerListenersResponse Int

The response's http status code.

DeleteLoadBalancer

deleteLoadBalancer_loadBalancerName :: Lens' DeleteLoadBalancer Text

The name of the load balancer.

deleteLoadBalancerResponse_httpStatus :: Lens' DeleteLoadBalancerResponse Int

The response's http status code.

AddTags

addTags_loadBalancerNames :: Lens' AddTags [Text]

The name of the load balancer. You can specify one load balancer only.

addTags_tags :: Lens' AddTags (NonEmpty Tag)

The tags.

addTagsResponse_httpStatus :: Lens' AddTagsResponse Int

The response's http status code.

ModifyLoadBalancerAttributes

modifyLoadBalancerAttributes_loadBalancerName :: Lens' ModifyLoadBalancerAttributes Text

The name of the load balancer.

modifyLoadBalancerAttributes_loadBalancerAttributes :: Lens' ModifyLoadBalancerAttributes LoadBalancerAttributes

The attributes for the load balancer.

modifyLoadBalancerAttributesResponse_loadBalancerAttributes :: Lens' ModifyLoadBalancerAttributesResponse (Maybe LoadBalancerAttributes)

Information about the load balancer attributes.

modifyLoadBalancerAttributesResponse_loadBalancerName :: Lens' ModifyLoadBalancerAttributesResponse (Maybe Text)

The name of the load balancer.

modifyLoadBalancerAttributesResponse_httpStatus :: Lens' ModifyLoadBalancerAttributesResponse Int

The response's http status code.

DescribeAccountLimits

describeAccountLimits_pageSize :: Lens' DescribeAccountLimits (Maybe Natural)

The maximum number of results to return with this call.

describeAccountLimits_marker :: Lens' DescribeAccountLimits (Maybe Text)

The marker for the next set of results. (You received this marker from a previous call.)

describeAccountLimitsResponse_nextMarker :: Lens' DescribeAccountLimitsResponse (Maybe Text)

The marker to use when requesting the next set of results. If there are no additional results, the string is empty.

describeAccountLimitsResponse_limits :: Lens' DescribeAccountLimitsResponse (Maybe [Limit])

Information about the limits.

describeAccountLimitsResponse_httpStatus :: Lens' DescribeAccountLimitsResponse Int

The response's http status code.

CreateLBCookieStickinessPolicy

createLBCookieStickinessPolicy_cookieExpirationPeriod :: Lens' CreateLBCookieStickinessPolicy (Maybe Integer)

The time period, in seconds, after which the cookie should be considered stale. If you do not specify this parameter, the default value is 0, which indicates that the sticky session should last for the duration of the browser session.

createLBCookieStickinessPolicy_loadBalancerName :: Lens' CreateLBCookieStickinessPolicy Text

The name of the load balancer.

createLBCookieStickinessPolicy_policyName :: Lens' CreateLBCookieStickinessPolicy Text

The name of the policy being created. Policy names must consist of alphanumeric characters and dashes (-). This name must be unique within the set of policies for this load balancer.

createLBCookieStickinessPolicyResponse_httpStatus :: Lens' CreateLBCookieStickinessPolicyResponse Int

The response's http status code.

AttachLoadBalancerToSubnets

attachLoadBalancerToSubnets_loadBalancerName :: Lens' AttachLoadBalancerToSubnets Text

The name of the load balancer.

attachLoadBalancerToSubnets_subnets :: Lens' AttachLoadBalancerToSubnets [Text]

The IDs of the subnets to add. You can add only one subnet per Availability Zone.

attachLoadBalancerToSubnetsResponse_subnets :: Lens' AttachLoadBalancerToSubnetsResponse (Maybe [Text])

The IDs of the subnets attached to the load balancer.

attachLoadBalancerToSubnetsResponse_httpStatus :: Lens' AttachLoadBalancerToSubnetsResponse Int

The response's http status code.

ApplySecurityGroupsToLoadBalancer

applySecurityGroupsToLoadBalancer_loadBalancerName :: Lens' ApplySecurityGroupsToLoadBalancer Text

The name of the load balancer.

applySecurityGroupsToLoadBalancer_securityGroups :: Lens' ApplySecurityGroupsToLoadBalancer [Text]

The IDs of the security groups to associate with the load balancer. Note that you cannot specify the name of the security group.

applySecurityGroupsToLoadBalancerResponse_securityGroups :: Lens' ApplySecurityGroupsToLoadBalancerResponse (Maybe [Text])

The IDs of the security groups associated with the load balancer.

applySecurityGroupsToLoadBalancerResponse_httpStatus :: Lens' ApplySecurityGroupsToLoadBalancerResponse Int

The response's http status code.

DescribeLoadBalancerPolicyTypes

describeLoadBalancerPolicyTypes_policyTypeNames :: Lens' DescribeLoadBalancerPolicyTypes (Maybe [Text])

The names of the policy types. If no names are specified, describes all policy types defined by Elastic Load Balancing.

describeLoadBalancerPolicyTypesResponse_policyTypeDescriptions :: Lens' DescribeLoadBalancerPolicyTypesResponse (Maybe [PolicyTypeDescription])

Information about the policy types.

describeLoadBalancerPolicyTypesResponse_httpStatus :: Lens' DescribeLoadBalancerPolicyTypesResponse Int

The response's http status code.

DescribeLoadBalancers

describeLoadBalancers_pageSize :: Lens' DescribeLoadBalancers (Maybe Natural)

The maximum number of results to return with this call (a number from 1 to 400). The default is 400.

describeLoadBalancers_loadBalancerNames :: Lens' DescribeLoadBalancers (Maybe [Text])

The names of the load balancers.

describeLoadBalancers_marker :: Lens' DescribeLoadBalancers (Maybe Text)

The marker for the next set of results. (You received this marker from a previous call.)

describeLoadBalancersResponse_loadBalancerDescriptions :: Lens' DescribeLoadBalancersResponse (Maybe [LoadBalancerDescription])

Information about the load balancers.

describeLoadBalancersResponse_nextMarker :: Lens' DescribeLoadBalancersResponse (Maybe Text)

The marker to use when requesting the next set of results. If there are no additional results, the string is empty.

describeLoadBalancersResponse_httpStatus :: Lens' DescribeLoadBalancersResponse Int

The response's http status code.

EnableAvailabilityZonesForLoadBalancer

enableAvailabilityZonesForLoadBalancer_loadBalancerName :: Lens' EnableAvailabilityZonesForLoadBalancer Text

The name of the load balancer.

enableAvailabilityZonesForLoadBalancer_availabilityZones :: Lens' EnableAvailabilityZonesForLoadBalancer [Text]

The Availability Zones. These must be in the same region as the load balancer.

enableAvailabilityZonesForLoadBalancerResponse_availabilityZones :: Lens' EnableAvailabilityZonesForLoadBalancerResponse (Maybe [Text])

The updated list of Availability Zones for the load balancer.

enableAvailabilityZonesForLoadBalancerResponse_httpStatus :: Lens' EnableAvailabilityZonesForLoadBalancerResponse Int

The response's http status code.

DisableAvailabilityZonesForLoadBalancer

disableAvailabilityZonesForLoadBalancer_loadBalancerName :: Lens' DisableAvailabilityZonesForLoadBalancer Text

The name of the load balancer.

disableAvailabilityZonesForLoadBalancer_availabilityZones :: Lens' DisableAvailabilityZonesForLoadBalancer [Text]

The Availability Zones.

disableAvailabilityZonesForLoadBalancerResponse_availabilityZones :: Lens' DisableAvailabilityZonesForLoadBalancerResponse (Maybe [Text])

The remaining Availability Zones for the load balancer.

disableAvailabilityZonesForLoadBalancerResponse_httpStatus :: Lens' DisableAvailabilityZonesForLoadBalancerResponse Int

The response's http status code.

DeleteLoadBalancerListeners

deleteLoadBalancerListeners_loadBalancerName :: Lens' DeleteLoadBalancerListeners Text

The name of the load balancer.

deleteLoadBalancerListeners_loadBalancerPorts :: Lens' DeleteLoadBalancerListeners [Int]

The client port numbers of the listeners.

deleteLoadBalancerListenersResponse_httpStatus :: Lens' DeleteLoadBalancerListenersResponse Int

The response's http status code.

CreateLoadBalancerPolicy

createLoadBalancerPolicy_policyAttributes :: Lens' CreateLoadBalancerPolicy (Maybe [PolicyAttribute])

The policy attributes.

createLoadBalancerPolicy_loadBalancerName :: Lens' CreateLoadBalancerPolicy Text

The name of the load balancer.

createLoadBalancerPolicy_policyName :: Lens' CreateLoadBalancerPolicy Text

The name of the load balancer policy to be created. This name must be unique within the set of policies for this load balancer.

createLoadBalancerPolicy_policyTypeName :: Lens' CreateLoadBalancerPolicy Text

The name of the base policy type. To get the list of policy types, use DescribeLoadBalancerPolicyTypes.

createLoadBalancerPolicyResponse_httpStatus :: Lens' CreateLoadBalancerPolicyResponse Int

The response's http status code.

DeleteLoadBalancerPolicy

deleteLoadBalancerPolicy_loadBalancerName :: Lens' DeleteLoadBalancerPolicy Text

The name of the load balancer.

deleteLoadBalancerPolicy_policyName :: Lens' DeleteLoadBalancerPolicy Text

The name of the policy.

deleteLoadBalancerPolicyResponse_httpStatus :: Lens' DeleteLoadBalancerPolicyResponse Int

The response's http status code.

DescribeLoadBalancerAttributes

describeLoadBalancerAttributes_loadBalancerName :: Lens' DescribeLoadBalancerAttributes Text

The name of the load balancer.

describeLoadBalancerAttributesResponse_loadBalancerAttributes :: Lens' DescribeLoadBalancerAttributesResponse (Maybe LoadBalancerAttributes)

Information about the load balancer attributes.

describeLoadBalancerAttributesResponse_httpStatus :: Lens' DescribeLoadBalancerAttributesResponse Int

The response's http status code.

DescribeInstanceHealth

describeInstanceHealth_instances :: Lens' DescribeInstanceHealth (Maybe [Instance])

The IDs of the instances.

describeInstanceHealth_loadBalancerName :: Lens' DescribeInstanceHealth Text

The name of the load balancer.

describeInstanceHealthResponse_instanceStates :: Lens' DescribeInstanceHealthResponse (Maybe [InstanceState])

Information about the health of the instances.

describeInstanceHealthResponse_httpStatus :: Lens' DescribeInstanceHealthResponse Int

The response's http status code.

CreateAppCookieStickinessPolicy

createAppCookieStickinessPolicy_loadBalancerName :: Lens' CreateAppCookieStickinessPolicy Text

The name of the load balancer.

createAppCookieStickinessPolicy_policyName :: Lens' CreateAppCookieStickinessPolicy Text

The name of the policy being created. Policy names must consist of alphanumeric characters and dashes (-). This name must be unique within the set of policies for this load balancer.

createAppCookieStickinessPolicy_cookieName :: Lens' CreateAppCookieStickinessPolicy Text

The name of the application cookie used for stickiness.

createAppCookieStickinessPolicyResponse_httpStatus :: Lens' CreateAppCookieStickinessPolicyResponse Int

The response's http status code.

ConfigureHealthCheck

configureHealthCheck_loadBalancerName :: Lens' ConfigureHealthCheck Text

The name of the load balancer.

configureHealthCheck_healthCheck :: Lens' ConfigureHealthCheck HealthCheck

The configuration information.

configureHealthCheckResponse_healthCheck :: Lens' ConfigureHealthCheckResponse (Maybe HealthCheck)

The updated health check.

configureHealthCheckResponse_httpStatus :: Lens' ConfigureHealthCheckResponse Int

The response's http status code.

Types

AccessLog

accessLog_s3BucketPrefix :: Lens' AccessLog (Maybe Text)

The logical hierarchy you created for your Amazon S3 bucket, for example my-bucket-prefix/prod. If the prefix is not provided, the log is placed at the root level of the bucket.

accessLog_emitInterval :: Lens' AccessLog (Maybe Int)

The interval for publishing the access logs. You can specify an interval of either 5 minutes or 60 minutes.

Default: 60 minutes

accessLog_s3BucketName :: Lens' AccessLog (Maybe Text)

The name of the Amazon S3 bucket where the access logs are stored.

accessLog_enabled :: Lens' AccessLog Bool

Specifies whether access logs are enabled for the load balancer.

AdditionalAttribute

additionalAttribute_key :: Lens' AdditionalAttribute (Maybe Text)

The name of the attribute.

The following attribute is supported.

elb.http.desyncmitigationmode - Determines how the load balancer handles requests that might pose a security risk to your application. The possible values are monitor, defensive, and strictest. The default is defensive.

additionalAttribute_value :: Lens' AdditionalAttribute (Maybe Text)

This value of the attribute.

AppCookieStickinessPolicy

appCookieStickinessPolicy_policyName :: Lens' AppCookieStickinessPolicy (Maybe Text)

The mnemonic name for the policy being created. The name must be unique within a set of policies for this load balancer.

appCookieStickinessPolicy_cookieName :: Lens' AppCookieStickinessPolicy (Maybe Text)

The name of the application cookie used for stickiness.

BackendServerDescription

backendServerDescription_instancePort :: Lens' BackendServerDescription (Maybe Natural)

The port on which the EC2 instance is listening.

backendServerDescription_policyNames :: Lens' BackendServerDescription (Maybe [Text])

The names of the policies enabled for the EC2 instance.

ConnectionDraining

connectionDraining_timeout :: Lens' ConnectionDraining (Maybe Int)

The maximum time, in seconds, to keep the existing connections open before deregistering the instances.

connectionDraining_enabled :: Lens' ConnectionDraining Bool

Specifies whether connection draining is enabled for the load balancer.

ConnectionSettings

connectionSettings_idleTimeout :: Lens' ConnectionSettings Natural

The time, in seconds, that the connection is allowed to be idle (no data has been sent over the connection) before it is closed by the load balancer.

CrossZoneLoadBalancing

crossZoneLoadBalancing_enabled :: Lens' CrossZoneLoadBalancing Bool

Specifies whether cross-zone load balancing is enabled for the load balancer.

HealthCheck

healthCheck_target :: Lens' HealthCheck Text

The instance being checked. The protocol is either TCP, HTTP, HTTPS, or SSL. The range of valid ports is one (1) through 65535.

TCP is the default, specified as a TCP: port pair, for example "TCP:5000". In this case, a health check simply attempts to open a TCP connection to the instance on the specified port. Failure to connect within the configured timeout is considered unhealthy.

SSL is also specified as SSL: port pair, for example, SSL:5000.

For HTTP/HTTPS, you must include a ping path in the string. HTTP is specified as a HTTP:port;/;PathToPing; grouping, for example "HTTP:80/weather/us/wa/seattle". In this case, a HTTP GET request is issued to the instance on the given port and path. Any answer other than "200 OK" within the timeout period is considered unhealthy.

The total length of the HTTP ping target must be 1024 16-bit Unicode characters or less.

healthCheck_interval :: Lens' HealthCheck Natural

The approximate interval, in seconds, between health checks of an individual instance.

healthCheck_timeout :: Lens' HealthCheck Natural

The amount of time, in seconds, during which no response means a failed health check.

This value must be less than the Interval value.

healthCheck_unhealthyThreshold :: Lens' HealthCheck Natural

The number of consecutive health check failures required before moving the instance to the Unhealthy state.

healthCheck_healthyThreshold :: Lens' HealthCheck Natural

The number of consecutive health checks successes required before moving the instance to the Healthy state.

Instance

instance_instanceId :: Lens' Instance (Maybe Text)

The instance ID.

InstanceState

instanceState_instanceId :: Lens' InstanceState (Maybe Text)

The ID of the instance.

instanceState_reasonCode :: Lens' InstanceState (Maybe Text)

Information about the cause of OutOfService instances. Specifically, whether the cause is Elastic Load Balancing or the instance.

Valid values: ELB | Instance | N/A

instanceState_state :: Lens' InstanceState (Maybe Text)

The current state of the instance.

Valid values: InService | OutOfService | Unknown

instanceState_description :: Lens' InstanceState (Maybe Text)

A description of the instance state. This string can contain one or more of the following messages.

N/A

A transient error occurred. Please try again later.

Instance has failed at least the UnhealthyThreshold number of health checks consecutively.

Instance has not passed the configured HealthyThreshold number of health checks consecutively.

Instance registration is still in progress.

Instance is in the EC2 Availability Zone for which LoadBalancer is not configured to route traffic to.

Instance is not currently registered with the LoadBalancer.

Instance deregistration currently in progress.

Disable Availability Zone is currently in progress.

Instance is in pending state.

Instance is in stopped state.

Instance is in terminated state.

LBCookieStickinessPolicy

lBCookieStickinessPolicy_policyName :: Lens' LBCookieStickinessPolicy (Maybe Text)

The name of the policy. This name must be unique within the set of policies for this load balancer.

lBCookieStickinessPolicy_cookieExpirationPeriod :: Lens' LBCookieStickinessPolicy (Maybe Integer)

The time period, in seconds, after which the cookie should be considered stale. If this parameter is not specified, the stickiness session lasts for the duration of the browser session.

Limit

limit_name :: Lens' Limit (Maybe Text)

The name of the limit. The possible values are:

classic-listeners

classic-load-balancers

classic-registered-instances

limit_max :: Lens' Limit (Maybe Text)

The maximum value of the limit.

Listener

listener_instanceProtocol :: Lens' Listener (Maybe Text)

The protocol to use for routing traffic to instances: HTTP, HTTPS, TCP, or SSL.

If the front-end protocol is TCP or SSL, the back-end protocol must be TCP or SSL. If the front-end protocol is HTTP or HTTPS, the back-end protocol must be HTTP or HTTPS.

If there is another listener with the same InstancePort whose InstanceProtocol is secure, (HTTPS or SSL), the listener's InstanceProtocol must also be secure.

If there is another listener with the same InstancePort whose InstanceProtocol is HTTP or TCP, the listener's InstanceProtocol must be HTTP or TCP.

listener_sSLCertificateId :: Lens' Listener (Maybe Text)

The Amazon Resource Name (ARN) of the server certificate.

listener_protocol :: Lens' Listener Text

The load balancer transport protocol to use for routing: HTTP, HTTPS, TCP, or SSL.

listener_loadBalancerPort :: Lens' Listener Int

The port on which the load balancer is listening. On EC2-VPC, you can specify any port from the range 1-65535. On EC2-Classic, you can specify any port from the following list: 25, 80, 443, 465, 587, 1024-65535.

listener_instancePort :: Lens' Listener Natural

The port on which the instance is listening.

ListenerDescription

listenerDescription_listener :: Lens' ListenerDescription (Maybe Listener)

The listener.

listenerDescription_policyNames :: Lens' ListenerDescription (Maybe [Text])

The policies. If there are no policies enabled, the list is empty.

LoadBalancerAttributes

loadBalancerAttributes_connectionDraining :: Lens' LoadBalancerAttributes (Maybe ConnectionDraining)

If enabled, the load balancer allows existing requests to complete before the load balancer shifts traffic away from a deregistered or unhealthy instance.

For more information, see Configure Connection Draining in the Classic Load Balancers Guide.

loadBalancerAttributes_additionalAttributes :: Lens' LoadBalancerAttributes (Maybe [AdditionalAttribute])

Any additional attributes.

loadBalancerAttributes_connectionSettings :: Lens' LoadBalancerAttributes (Maybe ConnectionSettings)

If enabled, the load balancer allows the connections to remain idle (no data is sent over the connection) for the specified duration.

By default, Elastic Load Balancing maintains a 60-second idle connection timeout for both front-end and back-end connections of your load balancer. For more information, see Configure Idle Connection Timeout in the Classic Load Balancers Guide.

loadBalancerAttributes_accessLog :: Lens' LoadBalancerAttributes (Maybe AccessLog)

If enabled, the load balancer captures detailed information of all requests and delivers the information to the Amazon S3 bucket that you specify.

For more information, see Enable Access Logs in the Classic Load Balancers Guide.

loadBalancerAttributes_crossZoneLoadBalancing :: Lens' LoadBalancerAttributes (Maybe CrossZoneLoadBalancing)

If enabled, the load balancer routes the request traffic evenly across all instances regardless of the Availability Zones.

For more information, see Configure Cross-Zone Load Balancing in the Classic Load Balancers Guide.

LoadBalancerDescription

loadBalancerDescription_canonicalHostedZoneNameID :: Lens' LoadBalancerDescription (Maybe Text)

The ID of the Amazon Route 53 hosted zone for the load balancer.

loadBalancerDescription_availabilityZones :: Lens' LoadBalancerDescription (Maybe [Text])

The Availability Zones for the load balancer.

loadBalancerDescription_backendServerDescriptions :: Lens' LoadBalancerDescription (Maybe [BackendServerDescription])

Information about your EC2 instances.

loadBalancerDescription_policies :: Lens' LoadBalancerDescription (Maybe Policies)

The policies defined for the load balancer.

loadBalancerDescription_scheme :: Lens' LoadBalancerDescription (Maybe Text)

The type of load balancer. Valid only for load balancers in a VPC.

If Scheme is internet-facing, the load balancer has a public DNS name that resolves to a public IP address.

If Scheme is internal, the load balancer has a public DNS name that resolves to a private IP address.

loadBalancerDescription_createdTime :: Lens' LoadBalancerDescription (Maybe UTCTime)

The date and time the load balancer was created.

loadBalancerDescription_instances :: Lens' LoadBalancerDescription (Maybe [Instance])

The IDs of the instances for the load balancer.

loadBalancerDescription_securityGroups :: Lens' LoadBalancerDescription (Maybe [Text])

The security groups for the load balancer. Valid only for load balancers in a VPC.

loadBalancerDescription_sourceSecurityGroup :: Lens' LoadBalancerDescription (Maybe SourceSecurityGroup)

The security group for the load balancer, which you can use as part of your inbound rules for your registered instances. To only allow traffic from load balancers, add a security group rule that specifies this source security group as the inbound source.

loadBalancerDescription_dNSName :: Lens' LoadBalancerDescription (Maybe Text)

The DNS name of the load balancer.

loadBalancerDescription_listenerDescriptions :: Lens' LoadBalancerDescription (Maybe [ListenerDescription])

The listeners for the load balancer.

loadBalancerDescription_subnets :: Lens' LoadBalancerDescription (Maybe [Text])

The IDs of the subnets for the load balancer.

loadBalancerDescription_vPCId :: Lens' LoadBalancerDescription (Maybe Text)

The ID of the VPC for the load balancer.

loadBalancerDescription_loadBalancerName :: Lens' LoadBalancerDescription (Maybe Text)

The name of the load balancer.

loadBalancerDescription_healthCheck :: Lens' LoadBalancerDescription (Maybe HealthCheck)

Information about the health checks conducted on the load balancer.

loadBalancerDescription_canonicalHostedZoneName :: Lens' LoadBalancerDescription (Maybe Text)

The DNS name of the load balancer.

For more information, see Configure a Custom Domain Name in the Classic Load Balancers Guide.

Policies

policies_appCookieStickinessPolicies :: Lens' Policies (Maybe [AppCookieStickinessPolicy])

The stickiness policies created using CreateAppCookieStickinessPolicy.

policies_lBCookieStickinessPolicies :: Lens' Policies (Maybe [LBCookieStickinessPolicy])

The stickiness policies created using CreateLBCookieStickinessPolicy.

policies_otherPolicies :: Lens' Policies (Maybe [Text])

The policies other than the stickiness policies.

PolicyAttribute

policyAttribute_attributeValue :: Lens' PolicyAttribute (Maybe Text)

The value of the attribute.

policyAttribute_attributeName :: Lens' PolicyAttribute (Maybe Text)

The name of the attribute.

PolicyAttributeDescription

policyAttributeDescription_attributeValue :: Lens' PolicyAttributeDescription (Maybe Text)

The value of the attribute.

policyAttributeDescription_attributeName :: Lens' PolicyAttributeDescription (Maybe Text)

The name of the attribute.

PolicyAttributeTypeDescription

policyAttributeTypeDescription_attributeType :: Lens' PolicyAttributeTypeDescription (Maybe Text)

The type of the attribute. For example, Boolean or Integer.

policyAttributeTypeDescription_attributeName :: Lens' PolicyAttributeTypeDescription (Maybe Text)

The name of the attribute.

policyAttributeTypeDescription_cardinality :: Lens' PolicyAttributeTypeDescription (Maybe Text)

The cardinality of the attribute.

Valid values:

ONE(1) : Single value required

ZERO_OR_ONE(0..1) : Up to one value is allowed

ZERO_OR_MORE(0..*) : Optional. Multiple values are allowed

ONE_OR_MORE(1..*0) : Required. Multiple values are allowed

policyAttributeTypeDescription_description :: Lens' PolicyAttributeTypeDescription (Maybe Text)

A description of the attribute.

policyAttributeTypeDescription_defaultValue :: Lens' PolicyAttributeTypeDescription (Maybe Text)

The default value of the attribute, if applicable.

PolicyDescription

policyDescription_policyName :: Lens' PolicyDescription (Maybe Text)

The name of the policy.

policyDescription_policyTypeName :: Lens' PolicyDescription (Maybe Text)

The name of the policy type.

policyDescription_policyAttributeDescriptions :: Lens' PolicyDescription (Maybe [PolicyAttributeDescription])

The policy attributes.

PolicyTypeDescription

policyTypeDescription_policyAttributeTypeDescriptions :: Lens' PolicyTypeDescription (Maybe [PolicyAttributeTypeDescription])

The description of the policy attributes associated with the policies defined by Elastic Load Balancing.

policyTypeDescription_policyTypeName :: Lens' PolicyTypeDescription (Maybe Text)

The name of the policy type.

policyTypeDescription_description :: Lens' PolicyTypeDescription (Maybe Text)

A description of the policy type.

SourceSecurityGroup

sourceSecurityGroup_ownerAlias :: Lens' SourceSecurityGroup (Maybe Text)

The owner of the security group.

sourceSecurityGroup_groupName :: Lens' SourceSecurityGroup (Maybe Text)

The name of the security group.

Tag

tag_value :: Lens' Tag (Maybe Text)

The value of the tag.

tag_key :: Lens' Tag Text

The key of the tag.

TagDescription

tagDescription_tags :: Lens' TagDescription (Maybe (NonEmpty Tag))

The tags.

tagDescription_loadBalancerName :: Lens' TagDescription (Maybe Text)

The name of the load balancer.

TagKeyOnly

tagKeyOnly_key :: Lens' TagKeyOnly (Maybe Text)

The name of the key.