Network.AWS.ELB.ModifyLoadBalancerAttributes

Description

Modifies the attributes of the specified load balancer.

You can modify the load balancer attributes, such as AccessLogs, ConnectionDraining, and CrossZoneLoadBalancing by either enabling or disabling them. Or, you can modify the load balancer attribute ConnectionSettings by specifying an idle connection timeout value for your load balancer.

For more information, see the following in the /Classic Load Balancers Guide/:

Cross-Zone Load Balancing

Connection Draining

Access Logs

Idle Connection Timeout

Creating a Request

data ModifyLoadBalancerAttributes

Contains the parameters for ModifyLoadBalancerAttributes.

See: newModifyLoadBalancerAttributes smart constructor.

Constructors

ModifyLoadBalancerAttributes'

Fields

loadBalancerName :: Text

The name of the load balancer.

loadBalancerAttributes :: LoadBalancerAttributes

The attributes for the load balancer.

newModifyLoadBalancerAttributes

Arguments

:: Text
$sel:loadBalancerName:ModifyLoadBalancerAttributes'

-> LoadBalancerAttributes
$sel:loadBalancerAttributes:ModifyLoadBalancerAttributes'

-> ModifyLoadBalancerAttributes

Create a value of ModifyLoadBalancerAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:loadBalancerName:ModifyLoadBalancerAttributes', modifyLoadBalancerAttributes_loadBalancerName - The name of the load balancer.

$sel:loadBalancerAttributes:ModifyLoadBalancerAttributes', modifyLoadBalancerAttributes_loadBalancerAttributes - The attributes for the load balancer.

Request Lenses

modifyLoadBalancerAttributes_loadBalancerName :: Lens' ModifyLoadBalancerAttributes Text

The name of the load balancer.

modifyLoadBalancerAttributes_loadBalancerAttributes :: Lens' ModifyLoadBalancerAttributes LoadBalancerAttributes

The attributes for the load balancer.

Destructuring the Response

data ModifyLoadBalancerAttributesResponse

Contains the output of ModifyLoadBalancerAttributes.

See: newModifyLoadBalancerAttributesResponse smart constructor.

Constructors

ModifyLoadBalancerAttributesResponse'

Fields

loadBalancerAttributes :: Maybe LoadBalancerAttributes

Information about the load balancer attributes.

loadBalancerName :: Maybe Text

The name of the load balancer.

httpStatus :: Int

The response's http status code.

newModifyLoadBalancerAttributesResponse

Arguments

:: Int
$sel:httpStatus:ModifyLoadBalancerAttributesResponse'

-> ModifyLoadBalancerAttributesResponse

Create a value of ModifyLoadBalancerAttributesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:loadBalancerAttributes:ModifyLoadBalancerAttributes', modifyLoadBalancerAttributesResponse_loadBalancerAttributes - Information about the load balancer attributes.

$sel:loadBalancerName:ModifyLoadBalancerAttributes', modifyLoadBalancerAttributesResponse_loadBalancerName - The name of the load balancer.

$sel:httpStatus:ModifyLoadBalancerAttributesResponse', modifyLoadBalancerAttributesResponse_httpStatus - The response's http status code.

Response Lenses

modifyLoadBalancerAttributesResponse_loadBalancerAttributes :: Lens' ModifyLoadBalancerAttributesResponse (Maybe LoadBalancerAttributes)

Information about the load balancer attributes.

modifyLoadBalancerAttributesResponse_loadBalancerName :: Lens' ModifyLoadBalancerAttributesResponse (Maybe Text)

The name of the load balancer.

modifyLoadBalancerAttributesResponse_httpStatus :: Lens' ModifyLoadBalancerAttributesResponse Int

The response's http status code.