Network.AWS.ELB.SetLoadBalancerPoliciesForBackendServer

Description

Replaces the set of policies associated with the specified port on which the EC2 instance is listening with a new set of policies. At this time, only the back-end server authentication policy type can be applied to the instance ports; this policy type is composed of multiple public key policies.

Each time you use SetLoadBalancerPoliciesForBackendServer to enable the policies, use the PolicyNames parameter to list the policies that you want to enable.

You can use DescribeLoadBalancers or DescribeLoadBalancerPolicies to verify that the policy is associated with the EC2 instance.

For more information about enabling back-end instance authentication, see Configure Back-end Instance Authentication in the Classic Load Balancers Guide. For more information about Proxy Protocol, see Configure Proxy Protocol Support in the Classic Load Balancers Guide.

Creating a Request

data SetLoadBalancerPoliciesForBackendServer

Contains the parameters for SetLoadBalancerPoliciesForBackendServer.

See: newSetLoadBalancerPoliciesForBackendServer smart constructor.

Constructors

SetLoadBalancerPoliciesForBackendServer'

Fields

loadBalancerName :: Text

The name of the load balancer.

instancePort :: Int

The port number associated with the EC2 instance.

policyNames :: [Text]

The names of the policies. If the list is empty, then all current polices are removed from the EC2 instance.

newSetLoadBalancerPoliciesForBackendServer

Arguments

:: Text
$sel:loadBalancerName:SetLoadBalancerPoliciesForBackendServer'

-> Int
$sel:instancePort:SetLoadBalancerPoliciesForBackendServer'

-> SetLoadBalancerPoliciesForBackendServer

Create a value of SetLoadBalancerPoliciesForBackendServer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:loadBalancerName:SetLoadBalancerPoliciesForBackendServer', setLoadBalancerPoliciesForBackendServer_loadBalancerName - The name of the load balancer.

$sel:instancePort:SetLoadBalancerPoliciesForBackendServer', setLoadBalancerPoliciesForBackendServer_instancePort - The port number associated with the EC2 instance.

$sel:policyNames:SetLoadBalancerPoliciesForBackendServer', setLoadBalancerPoliciesForBackendServer_policyNames - The names of the policies. If the list is empty, then all current polices are removed from the EC2 instance.

Request Lenses

setLoadBalancerPoliciesForBackendServer_loadBalancerName :: Lens' SetLoadBalancerPoliciesForBackendServer Text

The name of the load balancer.

setLoadBalancerPoliciesForBackendServer_instancePort :: Lens' SetLoadBalancerPoliciesForBackendServer Int

The port number associated with the EC2 instance.

setLoadBalancerPoliciesForBackendServer_policyNames :: Lens' SetLoadBalancerPoliciesForBackendServer [Text]

The names of the policies. If the list is empty, then all current polices are removed from the EC2 instance.

Destructuring the Response

data SetLoadBalancerPoliciesForBackendServerResponse

Contains the output of SetLoadBalancerPoliciesForBackendServer.

See: newSetLoadBalancerPoliciesForBackendServerResponse smart constructor.

Constructors

SetLoadBalancerPoliciesForBackendServerResponse'

Fields

httpStatus :: Int

The response's http status code.

newSetLoadBalancerPoliciesForBackendServerResponse

Arguments

:: Int
$sel:httpStatus:SetLoadBalancerPoliciesForBackendServerResponse'

-> SetLoadBalancerPoliciesForBackendServerResponse

Create a value of SetLoadBalancerPoliciesForBackendServerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:SetLoadBalancerPoliciesForBackendServerResponse', setLoadBalancerPoliciesForBackendServerResponse_httpStatus - The response's http status code.

Response Lenses

setLoadBalancerPoliciesForBackendServerResponse_httpStatus :: Lens' SetLoadBalancerPoliciesForBackendServerResponse Int

The response's http status code.