Network.AWS.ELB.Types

Description

Service Configuration

defaultService :: Service

API version 2012-06-01 of the Amazon Elastic Load Balancing SDK configuration.

Errors

_DuplicatePolicyNameException :: AsError a => Getting (First ServiceError) a ServiceError

A policy with the specified name already exists for this load balancer.

_InvalidConfigurationRequestException :: AsError a => Getting (First ServiceError) a ServiceError

The requested configuration change is not valid.

_DuplicateAccessPointNameException :: AsError a => Getting (First ServiceError) a ServiceError

The specified load balancer name already exists for this account.

_InvalidSecurityGroupException :: AsError a => Getting (First ServiceError) a ServiceError

One or more of the specified security groups do not exist.

_PolicyTypeNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

One or more of the specified policy types do not exist.

_AccessPointNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified load balancer does not exist.

_TooManyTagsException :: AsError a => Getting (First ServiceError) a ServiceError

The quota for the number of tags that can be assigned to a load balancer has been reached.

_DuplicateTagKeysException :: AsError a => Getting (First ServiceError) a ServiceError

A tag key was specified more than once.

_TooManyPoliciesException :: AsError a => Getting (First ServiceError) a ServiceError

The quota for the number of policies for this load balancer has been reached.

_DependencyThrottleException :: AsError a => Getting (First ServiceError) a ServiceError

A request made by Elastic Load Balancing to another service exceeds the maximum request rate permitted for your account.

_OperationNotPermittedException :: AsError a => Getting (First ServiceError) a ServiceError

This operation is not allowed.

_InvalidEndPointException :: AsError a => Getting (First ServiceError) a ServiceError

The specified endpoint is not valid.

_LoadBalancerAttributeNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified load balancer attribute does not exist.

_ListenerNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The load balancer does not have a listener configured at the specified port.

_SubnetNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

One or more of the specified subnets do not exist.

_PolicyNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

One or more of the specified policies do not exist.

_UnsupportedProtocolException :: AsError a => Getting (First ServiceError) a ServiceError

The specified protocol or signature version is not supported.

_CertificateNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified ARN does not refer to a valid SSL certificate in AWS Identity and Access Management (IAM) or AWS Certificate Manager (ACM). Note that if you recently uploaded the certificate to IAM, this error might indicate that the certificate is not fully available yet.

_DuplicateListenerException :: AsError a => Getting (First ServiceError) a ServiceError

A listener already exists for the specified load balancer name and port, but with a different instance port, protocol, or SSL certificate.

_InvalidSchemeException :: AsError a => Getting (First ServiceError) a ServiceError

The specified value for the schema is not valid. You can only specify a scheme for load balancers in a VPC.

_TooManyAccessPointsException :: AsError a => Getting (First ServiceError) a ServiceError

The quota for the number of load balancers has been reached.

_InvalidSubnetException :: AsError a => Getting (First ServiceError) a ServiceError

The specified VPC has no associated Internet gateway.

Re-exported Types

module Network.AWS.ELB.Internal

AccessLog

data AccessLog

Information about the AccessLog attribute.

See: newAccessLog smart constructor.

Constructors

AccessLog'

Fields

s3BucketPrefix :: Maybe Text

The logical hierarchy you created for your Amazon S3 bucket, for example my-bucket-prefix/prod. If the prefix is not provided, the log is placed at the root level of the bucket.

emitInterval :: Maybe Int

The interval for publishing the access logs. You can specify an interval of either 5 minutes or 60 minutes.

Default: 60 minutes

s3BucketName :: Maybe Text

The name of the Amazon S3 bucket where the access logs are stored.

enabled :: Bool

Specifies whether access logs are enabled for the load balancer.

newAccessLog

Arguments

:: Bool
$sel:enabled:AccessLog'

-> AccessLog

Create a value of AccessLog with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3BucketPrefix:AccessLog', accessLog_s3BucketPrefix - The logical hierarchy you created for your Amazon S3 bucket, for example my-bucket-prefix/prod. If the prefix is not provided, the log is placed at the root level of the bucket.

$sel:emitInterval:AccessLog', accessLog_emitInterval - The interval for publishing the access logs. You can specify an interval of either 5 minutes or 60 minutes.

Default: 60 minutes

$sel:s3BucketName:AccessLog', accessLog_s3BucketName - The name of the Amazon S3 bucket where the access logs are stored.

$sel:enabled:AccessLog', accessLog_enabled - Specifies whether access logs are enabled for the load balancer.

accessLog_s3BucketPrefix :: Lens' AccessLog (Maybe Text)

The logical hierarchy you created for your Amazon S3 bucket, for example my-bucket-prefix/prod. If the prefix is not provided, the log is placed at the root level of the bucket.

accessLog_emitInterval :: Lens' AccessLog (Maybe Int)

The interval for publishing the access logs. You can specify an interval of either 5 minutes or 60 minutes.

Default: 60 minutes

accessLog_s3BucketName :: Lens' AccessLog (Maybe Text)

The name of the Amazon S3 bucket where the access logs are stored.

accessLog_enabled :: Lens' AccessLog Bool

Specifies whether access logs are enabled for the load balancer.

AdditionalAttribute

data AdditionalAttribute

Information about additional load balancer attributes.

See: newAdditionalAttribute smart constructor.

Constructors

AdditionalAttribute'

Fields

key :: Maybe Text

The name of the attribute.

The following attribute is supported.

elb.http.desyncmitigationmode - Determines how the load balancer handles requests that might pose a security risk to your application. The possible values are monitor, defensive, and strictest. The default is defensive.

value :: Maybe Text

This value of the attribute.

newAdditionalAttribute :: AdditionalAttribute

Create a value of AdditionalAttribute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:AdditionalAttribute', additionalAttribute_key - The name of the attribute.

The following attribute is supported.

elb.http.desyncmitigationmode - Determines how the load balancer handles requests that might pose a security risk to your application. The possible values are monitor, defensive, and strictest. The default is defensive.

$sel:value:AdditionalAttribute', additionalAttribute_value - This value of the attribute.

additionalAttribute_key :: Lens' AdditionalAttribute (Maybe Text)

The name of the attribute.

The following attribute is supported.

elb.http.desyncmitigationmode - Determines how the load balancer handles requests that might pose a security risk to your application. The possible values are monitor, defensive, and strictest. The default is defensive.

additionalAttribute_value :: Lens' AdditionalAttribute (Maybe Text)

This value of the attribute.

AppCookieStickinessPolicy

data AppCookieStickinessPolicy

Information about a policy for application-controlled session stickiness.

See: newAppCookieStickinessPolicy smart constructor.

Constructors

AppCookieStickinessPolicy'

Fields

policyName :: Maybe Text

The mnemonic name for the policy being created. The name must be unique within a set of policies for this load balancer.

cookieName :: Maybe Text

The name of the application cookie used for stickiness.

newAppCookieStickinessPolicy :: AppCookieStickinessPolicy

Create a value of AppCookieStickinessPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policyName:AppCookieStickinessPolicy', appCookieStickinessPolicy_policyName - The mnemonic name for the policy being created. The name must be unique within a set of policies for this load balancer.

$sel:cookieName:AppCookieStickinessPolicy', appCookieStickinessPolicy_cookieName - The name of the application cookie used for stickiness.

appCookieStickinessPolicy_policyName :: Lens' AppCookieStickinessPolicy (Maybe Text)

The mnemonic name for the policy being created. The name must be unique within a set of policies for this load balancer.

appCookieStickinessPolicy_cookieName :: Lens' AppCookieStickinessPolicy (Maybe Text)

The name of the application cookie used for stickiness.

BackendServerDescription

data BackendServerDescription

Information about the configuration of an EC2 instance.

See: newBackendServerDescription smart constructor.

Constructors

BackendServerDescription'

Fields

instancePort :: Maybe Natural

The port on which the EC2 instance is listening.

policyNames :: Maybe [Text]

The names of the policies enabled for the EC2 instance.

newBackendServerDescription :: BackendServerDescription

Create a value of BackendServerDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instancePort:BackendServerDescription', backendServerDescription_instancePort - The port on which the EC2 instance is listening.

$sel:policyNames:BackendServerDescription', backendServerDescription_policyNames - The names of the policies enabled for the EC2 instance.

backendServerDescription_instancePort :: Lens' BackendServerDescription (Maybe Natural)

The port on which the EC2 instance is listening.

backendServerDescription_policyNames :: Lens' BackendServerDescription (Maybe [Text])

The names of the policies enabled for the EC2 instance.

ConnectionDraining

data ConnectionDraining

Information about the ConnectionDraining attribute.

See: newConnectionDraining smart constructor.

Constructors

ConnectionDraining'

Fields

timeout :: Maybe Int

The maximum time, in seconds, to keep the existing connections open before deregistering the instances.

enabled :: Bool

Specifies whether connection draining is enabled for the load balancer.

newConnectionDraining

Arguments

:: Bool
$sel:enabled:ConnectionDraining'

-> ConnectionDraining

Create a value of ConnectionDraining with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timeout:ConnectionDraining', connectionDraining_timeout - The maximum time, in seconds, to keep the existing connections open before deregistering the instances.

$sel:enabled:ConnectionDraining', connectionDraining_enabled - Specifies whether connection draining is enabled for the load balancer.

connectionDraining_timeout :: Lens' ConnectionDraining (Maybe Int)

The maximum time, in seconds, to keep the existing connections open before deregistering the instances.

connectionDraining_enabled :: Lens' ConnectionDraining Bool

Specifies whether connection draining is enabled for the load balancer.

ConnectionSettings

data ConnectionSettings

Information about the ConnectionSettings attribute.

See: newConnectionSettings smart constructor.

Constructors

ConnectionSettings'

Fields

idleTimeout :: Natural

The time, in seconds, that the connection is allowed to be idle (no data has been sent over the connection) before it is closed by the load balancer.

newConnectionSettings

Arguments

:: Natural
$sel:idleTimeout:ConnectionSettings'

-> ConnectionSettings

Create a value of ConnectionSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:idleTimeout:ConnectionSettings', connectionSettings_idleTimeout - The time, in seconds, that the connection is allowed to be idle (no data has been sent over the connection) before it is closed by the load balancer.

connectionSettings_idleTimeout :: Lens' ConnectionSettings Natural

The time, in seconds, that the connection is allowed to be idle (no data has been sent over the connection) before it is closed by the load balancer.

CrossZoneLoadBalancing

data CrossZoneLoadBalancing

Information about the CrossZoneLoadBalancing attribute.

See: newCrossZoneLoadBalancing smart constructor.

Constructors

CrossZoneLoadBalancing'

Fields

enabled :: Bool

Specifies whether cross-zone load balancing is enabled for the load balancer.

newCrossZoneLoadBalancing

Arguments

:: Bool
$sel:enabled:CrossZoneLoadBalancing'

-> CrossZoneLoadBalancing

Create a value of CrossZoneLoadBalancing with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:enabled:CrossZoneLoadBalancing', crossZoneLoadBalancing_enabled - Specifies whether cross-zone load balancing is enabled for the load balancer.

crossZoneLoadBalancing_enabled :: Lens' CrossZoneLoadBalancing Bool

Specifies whether cross-zone load balancing is enabled for the load balancer.

HealthCheck

data HealthCheck

Information about a health check.

See: newHealthCheck smart constructor.

Constructors

HealthCheck'

Fields

target :: Text

The instance being checked. The protocol is either TCP, HTTP, HTTPS, or SSL. The range of valid ports is one (1) through 65535.

TCP is the default, specified as a TCP: port pair, for example "TCP:5000". In this case, a health check simply attempts to open a TCP connection to the instance on the specified port. Failure to connect within the configured timeout is considered unhealthy.

SSL is also specified as SSL: port pair, for example, SSL:5000.

For HTTP/HTTPS, you must include a ping path in the string. HTTP is specified as a HTTP:port;/;PathToPing; grouping, for example "HTTP:80/weather/us/wa/seattle". In this case, a HTTP GET request is issued to the instance on the given port and path. Any answer other than "200 OK" within the timeout period is considered unhealthy.

The total length of the HTTP ping target must be 1024 16-bit Unicode characters or less.

interval :: Natural

The approximate interval, in seconds, between health checks of an individual instance.

timeout :: Natural

The amount of time, in seconds, during which no response means a failed health check.

This value must be less than the Interval value.

unhealthyThreshold :: Natural

The number of consecutive health check failures required before moving the instance to the Unhealthy state.

healthyThreshold :: Natural

The number of consecutive health checks successes required before moving the instance to the Healthy state.

newHealthCheck

Arguments

:: Text
$sel:target:HealthCheck'

-> Natural
$sel:interval:HealthCheck'

-> Natural
$sel:timeout:HealthCheck'

-> Natural
$sel:unhealthyThreshold:HealthCheck'

-> Natural
$sel:healthyThreshold:HealthCheck'

-> HealthCheck

Create a value of HealthCheck with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:target:HealthCheck', healthCheck_target - The instance being checked. The protocol is either TCP, HTTP, HTTPS, or SSL. The range of valid ports is one (1) through 65535.

TCP is the default, specified as a TCP: port pair, for example "TCP:5000". In this case, a health check simply attempts to open a TCP connection to the instance on the specified port. Failure to connect within the configured timeout is considered unhealthy.

SSL is also specified as SSL: port pair, for example, SSL:5000.

For HTTP/HTTPS, you must include a ping path in the string. HTTP is specified as a HTTP:port;/;PathToPing; grouping, for example "HTTP:80/weather/us/wa/seattle". In this case, a HTTP GET request is issued to the instance on the given port and path. Any answer other than "200 OK" within the timeout period is considered unhealthy.

The total length of the HTTP ping target must be 1024 16-bit Unicode characters or less.

$sel:interval:HealthCheck', healthCheck_interval - The approximate interval, in seconds, between health checks of an individual instance.

$sel:timeout:HealthCheck', healthCheck_timeout - The amount of time, in seconds, during which no response means a failed health check.

This value must be less than the Interval value.

$sel:unhealthyThreshold:HealthCheck', healthCheck_unhealthyThreshold - The number of consecutive health check failures required before moving the instance to the Unhealthy state.

$sel:healthyThreshold:HealthCheck', healthCheck_healthyThreshold - The number of consecutive health checks successes required before moving the instance to the Healthy state.

healthCheck_target :: Lens' HealthCheck Text

The instance being checked. The protocol is either TCP, HTTP, HTTPS, or SSL. The range of valid ports is one (1) through 65535.

TCP is the default, specified as a TCP: port pair, for example "TCP:5000". In this case, a health check simply attempts to open a TCP connection to the instance on the specified port. Failure to connect within the configured timeout is considered unhealthy.

SSL is also specified as SSL: port pair, for example, SSL:5000.

For HTTP/HTTPS, you must include a ping path in the string. HTTP is specified as a HTTP:port;/;PathToPing; grouping, for example "HTTP:80/weather/us/wa/seattle". In this case, a HTTP GET request is issued to the instance on the given port and path. Any answer other than "200 OK" within the timeout period is considered unhealthy.

The total length of the HTTP ping target must be 1024 16-bit Unicode characters or less.

healthCheck_interval :: Lens' HealthCheck Natural

The approximate interval, in seconds, between health checks of an individual instance.

healthCheck_timeout :: Lens' HealthCheck Natural

The amount of time, in seconds, during which no response means a failed health check.

This value must be less than the Interval value.

healthCheck_unhealthyThreshold :: Lens' HealthCheck Natural

The number of consecutive health check failures required before moving the instance to the Unhealthy state.

healthCheck_healthyThreshold :: Lens' HealthCheck Natural

The number of consecutive health checks successes required before moving the instance to the Healthy state.

Instance

data Instance

The ID of an EC2 instance.

See: newInstance smart constructor.

Constructors

Instance'

Fields

instanceId :: Maybe Text

The instance ID.

newInstance :: Instance

Create a value of Instance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceId:Instance', instance_instanceId - The instance ID.

instance_instanceId :: Lens' Instance (Maybe Text)

The instance ID.

InstanceState

data InstanceState

Information about the state of an EC2 instance.

See: newInstanceState smart constructor.

Constructors

InstanceState'

Fields

instanceId :: Maybe Text

The ID of the instance.

reasonCode :: Maybe Text

Information about the cause of OutOfService instances. Specifically, whether the cause is Elastic Load Balancing or the instance.

Valid values: ELB | Instance | N/A

state :: Maybe Text

The current state of the instance.

Valid values: InService | OutOfService | Unknown

description :: Maybe Text

A description of the instance state. This string can contain one or more of the following messages.

N/A

A transient error occurred. Please try again later.

Instance has failed at least the UnhealthyThreshold number of health checks consecutively.

Instance has not passed the configured HealthyThreshold number of health checks consecutively.

Instance registration is still in progress.

Instance is in the EC2 Availability Zone for which LoadBalancer is not configured to route traffic to.

Instance is not currently registered with the LoadBalancer.

Instance deregistration currently in progress.

Disable Availability Zone is currently in progress.

Instance is in pending state.

Instance is in stopped state.

Instance is in terminated state.

newInstanceState :: InstanceState

Create a value of InstanceState with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceId:InstanceState', instanceState_instanceId - The ID of the instance.

$sel:reasonCode:InstanceState', instanceState_reasonCode - Information about the cause of OutOfService instances. Specifically, whether the cause is Elastic Load Balancing or the instance.

Valid values: ELB | Instance | N/A

$sel:state:InstanceState', instanceState_state - The current state of the instance.

Valid values: InService | OutOfService | Unknown

$sel:description:InstanceState', instanceState_description - A description of the instance state. This string can contain one or more of the following messages.

N/A

A transient error occurred. Please try again later.

Instance has failed at least the UnhealthyThreshold number of health checks consecutively.

Instance has not passed the configured HealthyThreshold number of health checks consecutively.

Instance registration is still in progress.

Instance is in the EC2 Availability Zone for which LoadBalancer is not configured to route traffic to.

Instance is not currently registered with the LoadBalancer.

Instance deregistration currently in progress.

Disable Availability Zone is currently in progress.

Instance is in pending state.

Instance is in stopped state.

Instance is in terminated state.

instanceState_instanceId :: Lens' InstanceState (Maybe Text)

The ID of the instance.

instanceState_reasonCode :: Lens' InstanceState (Maybe Text)

Information about the cause of OutOfService instances. Specifically, whether the cause is Elastic Load Balancing or the instance.

Valid values: ELB | Instance | N/A

instanceState_state :: Lens' InstanceState (Maybe Text)

The current state of the instance.

Valid values: InService | OutOfService | Unknown

instanceState_description :: Lens' InstanceState (Maybe Text)

A description of the instance state. This string can contain one or more of the following messages.

N/A

A transient error occurred. Please try again later.

Instance has failed at least the UnhealthyThreshold number of health checks consecutively.

Instance has not passed the configured HealthyThreshold number of health checks consecutively.

Instance registration is still in progress.

Instance is in the EC2 Availability Zone for which LoadBalancer is not configured to route traffic to.

Instance is not currently registered with the LoadBalancer.

Instance deregistration currently in progress.

Disable Availability Zone is currently in progress.

Instance is in pending state.

Instance is in stopped state.

Instance is in terminated state.

LBCookieStickinessPolicy

data LBCookieStickinessPolicy

Information about a policy for duration-based session stickiness.

See: newLBCookieStickinessPolicy smart constructor.

Constructors

LBCookieStickinessPolicy'

Fields

policyName :: Maybe Text

The name of the policy. This name must be unique within the set of policies for this load balancer.

cookieExpirationPeriod :: Maybe Integer

The time period, in seconds, after which the cookie should be considered stale. If this parameter is not specified, the stickiness session lasts for the duration of the browser session.

newLBCookieStickinessPolicy :: LBCookieStickinessPolicy

Create a value of LBCookieStickinessPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policyName:LBCookieStickinessPolicy', lBCookieStickinessPolicy_policyName - The name of the policy. This name must be unique within the set of policies for this load balancer.

$sel:cookieExpirationPeriod:LBCookieStickinessPolicy', lBCookieStickinessPolicy_cookieExpirationPeriod - The time period, in seconds, after which the cookie should be considered stale. If this parameter is not specified, the stickiness session lasts for the duration of the browser session.

lBCookieStickinessPolicy_policyName :: Lens' LBCookieStickinessPolicy (Maybe Text)

The name of the policy. This name must be unique within the set of policies for this load balancer.

lBCookieStickinessPolicy_cookieExpirationPeriod :: Lens' LBCookieStickinessPolicy (Maybe Integer)

The time period, in seconds, after which the cookie should be considered stale. If this parameter is not specified, the stickiness session lasts for the duration of the browser session.

Limit

data Limit

Information about an Elastic Load Balancing resource limit for your AWS account.

See: newLimit smart constructor.

Constructors

Limit'

Fields

name :: Maybe Text

The name of the limit. The possible values are:

classic-listeners

classic-load-balancers

classic-registered-instances

max :: Maybe Text

The maximum value of the limit.

newLimit :: Limit

Create a value of Limit with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:Limit', limit_name - The name of the limit. The possible values are:

classic-listeners

classic-load-balancers

classic-registered-instances

$sel:max:Limit', limit_max - The maximum value of the limit.

limit_name :: Lens' Limit (Maybe Text)

The name of the limit. The possible values are:

classic-listeners

classic-load-balancers

classic-registered-instances

limit_max :: Lens' Limit (Maybe Text)

The maximum value of the limit.

Listener

data Listener

Information about a listener.

For information about the protocols and the ports supported by Elastic Load Balancing, see Listeners for Your Classic Load Balancer in the Classic Load Balancers Guide.

See: newListener smart constructor.

Constructors

Listener'

Fields

instanceProtocol :: Maybe Text

The protocol to use for routing traffic to instances: HTTP, HTTPS, TCP, or SSL.

If the front-end protocol is TCP or SSL, the back-end protocol must be TCP or SSL. If the front-end protocol is HTTP or HTTPS, the back-end protocol must be HTTP or HTTPS.

If there is another listener with the same InstancePort whose InstanceProtocol is secure, (HTTPS or SSL), the listener's InstanceProtocol must also be secure.

If there is another listener with the same InstancePort whose InstanceProtocol is HTTP or TCP, the listener's InstanceProtocol must be HTTP or TCP.

sSLCertificateId :: Maybe Text

The Amazon Resource Name (ARN) of the server certificate.

protocol :: Text

The load balancer transport protocol to use for routing: HTTP, HTTPS, TCP, or SSL.

loadBalancerPort :: Int

The port on which the load balancer is listening. On EC2-VPC, you can specify any port from the range 1-65535. On EC2-Classic, you can specify any port from the following list: 25, 80, 443, 465, 587, 1024-65535.

instancePort :: Natural

The port on which the instance is listening.

newListener

Arguments

:: Text
$sel:protocol:Listener'

-> Int
$sel:loadBalancerPort:Listener'

-> Natural
$sel:instancePort:Listener'

-> Listener

Create a value of Listener with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceProtocol:Listener', listener_instanceProtocol - The protocol to use for routing traffic to instances: HTTP, HTTPS, TCP, or SSL.

If the front-end protocol is TCP or SSL, the back-end protocol must be TCP or SSL. If the front-end protocol is HTTP or HTTPS, the back-end protocol must be HTTP or HTTPS.

If there is another listener with the same InstancePort whose InstanceProtocol is secure, (HTTPS or SSL), the listener's InstanceProtocol must also be secure.

If there is another listener with the same InstancePort whose InstanceProtocol is HTTP or TCP, the listener's InstanceProtocol must be HTTP or TCP.

$sel:sSLCertificateId:Listener', listener_sSLCertificateId - The Amazon Resource Name (ARN) of the server certificate.

$sel:protocol:Listener', listener_protocol - The load balancer transport protocol to use for routing: HTTP, HTTPS, TCP, or SSL.

$sel:loadBalancerPort:Listener', listener_loadBalancerPort - The port on which the load balancer is listening. On EC2-VPC, you can specify any port from the range 1-65535. On EC2-Classic, you can specify any port from the following list: 25, 80, 443, 465, 587, 1024-65535.

$sel:instancePort:Listener', listener_instancePort - The port on which the instance is listening.

listener_instanceProtocol :: Lens' Listener (Maybe Text)

The protocol to use for routing traffic to instances: HTTP, HTTPS, TCP, or SSL.

If the front-end protocol is TCP or SSL, the back-end protocol must be TCP or SSL. If the front-end protocol is HTTP or HTTPS, the back-end protocol must be HTTP or HTTPS.

If there is another listener with the same InstancePort whose InstanceProtocol is secure, (HTTPS or SSL), the listener's InstanceProtocol must also be secure.

If there is another listener with the same InstancePort whose InstanceProtocol is HTTP or TCP, the listener's InstanceProtocol must be HTTP or TCP.

listener_sSLCertificateId :: Lens' Listener (Maybe Text)

The Amazon Resource Name (ARN) of the server certificate.

listener_protocol :: Lens' Listener Text

The load balancer transport protocol to use for routing: HTTP, HTTPS, TCP, or SSL.

listener_loadBalancerPort :: Lens' Listener Int

The port on which the load balancer is listening. On EC2-VPC, you can specify any port from the range 1-65535. On EC2-Classic, you can specify any port from the following list: 25, 80, 443, 465, 587, 1024-65535.

listener_instancePort :: Lens' Listener Natural

The port on which the instance is listening.

ListenerDescription

data ListenerDescription

The policies enabled for a listener.

See: newListenerDescription smart constructor.

Constructors

ListenerDescription'

Fields

listener :: Maybe Listener

The listener.

policyNames :: Maybe [Text]

The policies. If there are no policies enabled, the list is empty.

newListenerDescription :: ListenerDescription

Create a value of ListenerDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:listener:ListenerDescription', listenerDescription_listener - The listener.

$sel:policyNames:ListenerDescription', listenerDescription_policyNames - The policies. If there are no policies enabled, the list is empty.

listenerDescription_listener :: Lens' ListenerDescription (Maybe Listener)

The listener.

listenerDescription_policyNames :: Lens' ListenerDescription (Maybe [Text])

The policies. If there are no policies enabled, the list is empty.

LoadBalancerAttributes

data LoadBalancerAttributes

The attributes for a load balancer.

See: newLoadBalancerAttributes smart constructor.

Constructors

LoadBalancerAttributes'

Fields

connectionDraining :: Maybe ConnectionDraining

If enabled, the load balancer allows existing requests to complete before the load balancer shifts traffic away from a deregistered or unhealthy instance.

For more information, see Configure Connection Draining in the Classic Load Balancers Guide.

additionalAttributes :: Maybe [AdditionalAttribute]

Any additional attributes.

connectionSettings :: Maybe ConnectionSettings

If enabled, the load balancer allows the connections to remain idle (no data is sent over the connection) for the specified duration.

By default, Elastic Load Balancing maintains a 60-second idle connection timeout for both front-end and back-end connections of your load balancer. For more information, see Configure Idle Connection Timeout in the Classic Load Balancers Guide.

accessLog :: Maybe AccessLog

If enabled, the load balancer captures detailed information of all requests and delivers the information to the Amazon S3 bucket that you specify.

For more information, see Enable Access Logs in the Classic Load Balancers Guide.

crossZoneLoadBalancing :: Maybe CrossZoneLoadBalancing

If enabled, the load balancer routes the request traffic evenly across all instances regardless of the Availability Zones.

For more information, see Configure Cross-Zone Load Balancing in the Classic Load Balancers Guide.

newLoadBalancerAttributes :: LoadBalancerAttributes

Create a value of LoadBalancerAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:connectionDraining:LoadBalancerAttributes', loadBalancerAttributes_connectionDraining - If enabled, the load balancer allows existing requests to complete before the load balancer shifts traffic away from a deregistered or unhealthy instance.

For more information, see Configure Connection Draining in the Classic Load Balancers Guide.

$sel:additionalAttributes:LoadBalancerAttributes', loadBalancerAttributes_additionalAttributes - Any additional attributes.

$sel:connectionSettings:LoadBalancerAttributes', loadBalancerAttributes_connectionSettings - If enabled, the load balancer allows the connections to remain idle (no data is sent over the connection) for the specified duration.

By default, Elastic Load Balancing maintains a 60-second idle connection timeout for both front-end and back-end connections of your load balancer. For more information, see Configure Idle Connection Timeout in the Classic Load Balancers Guide.

$sel:accessLog:LoadBalancerAttributes', loadBalancerAttributes_accessLog - If enabled, the load balancer captures detailed information of all requests and delivers the information to the Amazon S3 bucket that you specify.

For more information, see Enable Access Logs in the Classic Load Balancers Guide.

$sel:crossZoneLoadBalancing:LoadBalancerAttributes', loadBalancerAttributes_crossZoneLoadBalancing - If enabled, the load balancer routes the request traffic evenly across all instances regardless of the Availability Zones.

For more information, see Configure Cross-Zone Load Balancing in the Classic Load Balancers Guide.

loadBalancerAttributes_connectionDraining :: Lens' LoadBalancerAttributes (Maybe ConnectionDraining)

If enabled, the load balancer allows existing requests to complete before the load balancer shifts traffic away from a deregistered or unhealthy instance.

For more information, see Configure Connection Draining in the Classic Load Balancers Guide.

loadBalancerAttributes_additionalAttributes :: Lens' LoadBalancerAttributes (Maybe [AdditionalAttribute])

Any additional attributes.

loadBalancerAttributes_connectionSettings :: Lens' LoadBalancerAttributes (Maybe ConnectionSettings)

If enabled, the load balancer allows the connections to remain idle (no data is sent over the connection) for the specified duration.

By default, Elastic Load Balancing maintains a 60-second idle connection timeout for both front-end and back-end connections of your load balancer. For more information, see Configure Idle Connection Timeout in the Classic Load Balancers Guide.

loadBalancerAttributes_accessLog :: Lens' LoadBalancerAttributes (Maybe AccessLog)

If enabled, the load balancer captures detailed information of all requests and delivers the information to the Amazon S3 bucket that you specify.

For more information, see Enable Access Logs in the Classic Load Balancers Guide.

loadBalancerAttributes_crossZoneLoadBalancing :: Lens' LoadBalancerAttributes (Maybe CrossZoneLoadBalancing)

If enabled, the load balancer routes the request traffic evenly across all instances regardless of the Availability Zones.

For more information, see Configure Cross-Zone Load Balancing in the Classic Load Balancers Guide.

LoadBalancerDescription

data LoadBalancerDescription

Information about a load balancer.

See: newLoadBalancerDescription smart constructor.

Constructors

LoadBalancerDescription'

Fields

canonicalHostedZoneNameID :: Maybe Text

The ID of the Amazon Route 53 hosted zone for the load balancer.

availabilityZones :: Maybe [Text]

The Availability Zones for the load balancer.

backendServerDescriptions :: Maybe [BackendServerDescription]

Information about your EC2 instances.

policies :: Maybe Policies

The policies defined for the load balancer.

scheme :: Maybe Text

The type of load balancer. Valid only for load balancers in a VPC.

If Scheme is internet-facing, the load balancer has a public DNS name that resolves to a public IP address.

If Scheme is internal, the load balancer has a public DNS name that resolves to a private IP address.

createdTime :: Maybe ISO8601

The date and time the load balancer was created.

instances :: Maybe [Instance]

The IDs of the instances for the load balancer.

securityGroups :: Maybe [Text]

The security groups for the load balancer. Valid only for load balancers in a VPC.

sourceSecurityGroup :: Maybe SourceSecurityGroup

The security group for the load balancer, which you can use as part of your inbound rules for your registered instances. To only allow traffic from load balancers, add a security group rule that specifies this source security group as the inbound source.

dNSName :: Maybe Text

The DNS name of the load balancer.

listenerDescriptions :: Maybe [ListenerDescription]

The listeners for the load balancer.

subnets :: Maybe [Text]

The IDs of the subnets for the load balancer.

vPCId :: Maybe Text

The ID of the VPC for the load balancer.

loadBalancerName :: Maybe Text

The name of the load balancer.

healthCheck :: Maybe HealthCheck

Information about the health checks conducted on the load balancer.

canonicalHostedZoneName :: Maybe Text

The DNS name of the load balancer.

For more information, see Configure a Custom Domain Name in the Classic Load Balancers Guide.

newLoadBalancerDescription :: LoadBalancerDescription

Create a value of LoadBalancerDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:canonicalHostedZoneNameID:LoadBalancerDescription', loadBalancerDescription_canonicalHostedZoneNameID - The ID of the Amazon Route 53 hosted zone for the load balancer.

$sel:availabilityZones:LoadBalancerDescription', loadBalancerDescription_availabilityZones - The Availability Zones for the load balancer.

$sel:backendServerDescriptions:LoadBalancerDescription', loadBalancerDescription_backendServerDescriptions - Information about your EC2 instances.

$sel:policies:LoadBalancerDescription', loadBalancerDescription_policies - The policies defined for the load balancer.

$sel:scheme:LoadBalancerDescription', loadBalancerDescription_scheme - The type of load balancer. Valid only for load balancers in a VPC.

If Scheme is internet-facing, the load balancer has a public DNS name that resolves to a public IP address.

If Scheme is internal, the load balancer has a public DNS name that resolves to a private IP address.

$sel:createdTime:LoadBalancerDescription', loadBalancerDescription_createdTime - The date and time the load balancer was created.

$sel:instances:LoadBalancerDescription', loadBalancerDescription_instances - The IDs of the instances for the load balancer.

$sel:securityGroups:LoadBalancerDescription', loadBalancerDescription_securityGroups - The security groups for the load balancer. Valid only for load balancers in a VPC.

$sel:sourceSecurityGroup:LoadBalancerDescription', loadBalancerDescription_sourceSecurityGroup - The security group for the load balancer, which you can use as part of your inbound rules for your registered instances. To only allow traffic from load balancers, add a security group rule that specifies this source security group as the inbound source.

$sel:dNSName:LoadBalancerDescription', loadBalancerDescription_dNSName - The DNS name of the load balancer.

$sel:listenerDescriptions:LoadBalancerDescription', loadBalancerDescription_listenerDescriptions - The listeners for the load balancer.

$sel:subnets:LoadBalancerDescription', loadBalancerDescription_subnets - The IDs of the subnets for the load balancer.

$sel:vPCId:LoadBalancerDescription', loadBalancerDescription_vPCId - The ID of the VPC for the load balancer.

$sel:loadBalancerName:LoadBalancerDescription', loadBalancerDescription_loadBalancerName - The name of the load balancer.

$sel:healthCheck:LoadBalancerDescription', loadBalancerDescription_healthCheck - Information about the health checks conducted on the load balancer.

$sel:canonicalHostedZoneName:LoadBalancerDescription', loadBalancerDescription_canonicalHostedZoneName - The DNS name of the load balancer.

For more information, see Configure a Custom Domain Name in the Classic Load Balancers Guide.

loadBalancerDescription_canonicalHostedZoneNameID :: Lens' LoadBalancerDescription (Maybe Text)

The ID of the Amazon Route 53 hosted zone for the load balancer.

loadBalancerDescription_availabilityZones :: Lens' LoadBalancerDescription (Maybe [Text])

The Availability Zones for the load balancer.

loadBalancerDescription_backendServerDescriptions :: Lens' LoadBalancerDescription (Maybe [BackendServerDescription])

Information about your EC2 instances.

loadBalancerDescription_policies :: Lens' LoadBalancerDescription (Maybe Policies)

The policies defined for the load balancer.

loadBalancerDescription_scheme :: Lens' LoadBalancerDescription (Maybe Text)

The type of load balancer. Valid only for load balancers in a VPC.

If Scheme is internet-facing, the load balancer has a public DNS name that resolves to a public IP address.

If Scheme is internal, the load balancer has a public DNS name that resolves to a private IP address.

loadBalancerDescription_createdTime :: Lens' LoadBalancerDescription (Maybe UTCTime)

The date and time the load balancer was created.

loadBalancerDescription_instances :: Lens' LoadBalancerDescription (Maybe [Instance])

The IDs of the instances for the load balancer.

loadBalancerDescription_securityGroups :: Lens' LoadBalancerDescription (Maybe [Text])

The security groups for the load balancer. Valid only for load balancers in a VPC.

loadBalancerDescription_sourceSecurityGroup :: Lens' LoadBalancerDescription (Maybe SourceSecurityGroup)

The security group for the load balancer, which you can use as part of your inbound rules for your registered instances. To only allow traffic from load balancers, add a security group rule that specifies this source security group as the inbound source.

loadBalancerDescription_dNSName :: Lens' LoadBalancerDescription (Maybe Text)

The DNS name of the load balancer.

loadBalancerDescription_listenerDescriptions :: Lens' LoadBalancerDescription (Maybe [ListenerDescription])

The listeners for the load balancer.

loadBalancerDescription_subnets :: Lens' LoadBalancerDescription (Maybe [Text])

The IDs of the subnets for the load balancer.

loadBalancerDescription_vPCId :: Lens' LoadBalancerDescription (Maybe Text)

The ID of the VPC for the load balancer.

loadBalancerDescription_loadBalancerName :: Lens' LoadBalancerDescription (Maybe Text)

The name of the load balancer.

loadBalancerDescription_healthCheck :: Lens' LoadBalancerDescription (Maybe HealthCheck)

Information about the health checks conducted on the load balancer.

loadBalancerDescription_canonicalHostedZoneName :: Lens' LoadBalancerDescription (Maybe Text)

The DNS name of the load balancer.

For more information, see Configure a Custom Domain Name in the Classic Load Balancers Guide.

Policies

data Policies

The policies for a load balancer.

See: newPolicies smart constructor.

Constructors

Policies'

Fields

appCookieStickinessPolicies :: Maybe [AppCookieStickinessPolicy]

The stickiness policies created using CreateAppCookieStickinessPolicy.

lBCookieStickinessPolicies :: Maybe [LBCookieStickinessPolicy]

The stickiness policies created using CreateLBCookieStickinessPolicy.

otherPolicies :: Maybe [Text]

The policies other than the stickiness policies.

newPolicies :: Policies

Create a value of Policies with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appCookieStickinessPolicies:Policies', policies_appCookieStickinessPolicies - The stickiness policies created using CreateAppCookieStickinessPolicy.

$sel:lBCookieStickinessPolicies:Policies', policies_lBCookieStickinessPolicies - The stickiness policies created using CreateLBCookieStickinessPolicy.

$sel:otherPolicies:Policies', policies_otherPolicies - The policies other than the stickiness policies.

policies_appCookieStickinessPolicies :: Lens' Policies (Maybe [AppCookieStickinessPolicy])

The stickiness policies created using CreateAppCookieStickinessPolicy.

policies_lBCookieStickinessPolicies :: Lens' Policies (Maybe [LBCookieStickinessPolicy])

The stickiness policies created using CreateLBCookieStickinessPolicy.

policies_otherPolicies :: Lens' Policies (Maybe [Text])

The policies other than the stickiness policies.

PolicyAttribute

data PolicyAttribute

Information about a policy attribute.

See: newPolicyAttribute smart constructor.

Constructors

PolicyAttribute'

Fields

attributeValue :: Maybe Text

The value of the attribute.

attributeName :: Maybe Text

The name of the attribute.

newPolicyAttribute :: PolicyAttribute

Create a value of PolicyAttribute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributeValue:PolicyAttribute', policyAttribute_attributeValue - The value of the attribute.

$sel:attributeName:PolicyAttribute', policyAttribute_attributeName - The name of the attribute.

policyAttribute_attributeValue :: Lens' PolicyAttribute (Maybe Text)

The value of the attribute.

policyAttribute_attributeName :: Lens' PolicyAttribute (Maybe Text)

The name of the attribute.

PolicyAttributeDescription

data PolicyAttributeDescription

Information about a policy attribute.

See: newPolicyAttributeDescription smart constructor.

Constructors

PolicyAttributeDescription'

Fields

attributeValue :: Maybe Text

The value of the attribute.

attributeName :: Maybe Text

The name of the attribute.

newPolicyAttributeDescription :: PolicyAttributeDescription

Create a value of PolicyAttributeDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributeValue:PolicyAttributeDescription', policyAttributeDescription_attributeValue - The value of the attribute.

$sel:attributeName:PolicyAttributeDescription', policyAttributeDescription_attributeName - The name of the attribute.

policyAttributeDescription_attributeValue :: Lens' PolicyAttributeDescription (Maybe Text)

The value of the attribute.

policyAttributeDescription_attributeName :: Lens' PolicyAttributeDescription (Maybe Text)

The name of the attribute.

PolicyAttributeTypeDescription

data PolicyAttributeTypeDescription

Information about a policy attribute type.

See: newPolicyAttributeTypeDescription smart constructor.

Constructors

PolicyAttributeTypeDescription'

Fields

attributeType :: Maybe Text

The type of the attribute. For example, Boolean or Integer.

attributeName :: Maybe Text

The name of the attribute.

cardinality :: Maybe Text

The cardinality of the attribute.

Valid values:

ONE(1) : Single value required

ZERO_OR_ONE(0..1) : Up to one value is allowed

ZERO_OR_MORE(0..*) : Optional. Multiple values are allowed

ONE_OR_MORE(1..*0) : Required. Multiple values are allowed

description :: Maybe Text

A description of the attribute.

defaultValue :: Maybe Text

The default value of the attribute, if applicable.

newPolicyAttributeTypeDescription :: PolicyAttributeTypeDescription

Create a value of PolicyAttributeTypeDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributeType:PolicyAttributeTypeDescription', policyAttributeTypeDescription_attributeType - The type of the attribute. For example, Boolean or Integer.

$sel:attributeName:PolicyAttributeTypeDescription', policyAttributeTypeDescription_attributeName - The name of the attribute.

$sel:cardinality:PolicyAttributeTypeDescription', policyAttributeTypeDescription_cardinality - The cardinality of the attribute.

Valid values:

ONE(1) : Single value required

ZERO_OR_ONE(0..1) : Up to one value is allowed

ZERO_OR_MORE(0..*) : Optional. Multiple values are allowed

ONE_OR_MORE(1..*0) : Required. Multiple values are allowed

$sel:description:PolicyAttributeTypeDescription', policyAttributeTypeDescription_description - A description of the attribute.

$sel:defaultValue:PolicyAttributeTypeDescription', policyAttributeTypeDescription_defaultValue - The default value of the attribute, if applicable.

policyAttributeTypeDescription_attributeType :: Lens' PolicyAttributeTypeDescription (Maybe Text)

The type of the attribute. For example, Boolean or Integer.

policyAttributeTypeDescription_attributeName :: Lens' PolicyAttributeTypeDescription (Maybe Text)

The name of the attribute.

policyAttributeTypeDescription_cardinality :: Lens' PolicyAttributeTypeDescription (Maybe Text)

The cardinality of the attribute.

Valid values:

ONE(1) : Single value required

ZERO_OR_ONE(0..1) : Up to one value is allowed

ZERO_OR_MORE(0..*) : Optional. Multiple values are allowed

ONE_OR_MORE(1..*0) : Required. Multiple values are allowed

policyAttributeTypeDescription_description :: Lens' PolicyAttributeTypeDescription (Maybe Text)

A description of the attribute.

policyAttributeTypeDescription_defaultValue :: Lens' PolicyAttributeTypeDescription (Maybe Text)

The default value of the attribute, if applicable.

PolicyDescription

data PolicyDescription

Information about a policy.

See: newPolicyDescription smart constructor.

Constructors

PolicyDescription'

Fields

policyName :: Maybe Text

The name of the policy.

policyTypeName :: Maybe Text

The name of the policy type.

policyAttributeDescriptions :: Maybe [PolicyAttributeDescription]

The policy attributes.

newPolicyDescription :: PolicyDescription

Create a value of PolicyDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policyName:PolicyDescription', policyDescription_policyName - The name of the policy.

$sel:policyTypeName:PolicyDescription', policyDescription_policyTypeName - The name of the policy type.

$sel:policyAttributeDescriptions:PolicyDescription', policyDescription_policyAttributeDescriptions - The policy attributes.

policyDescription_policyName :: Lens' PolicyDescription (Maybe Text)

The name of the policy.

policyDescription_policyTypeName :: Lens' PolicyDescription (Maybe Text)

The name of the policy type.

policyDescription_policyAttributeDescriptions :: Lens' PolicyDescription (Maybe [PolicyAttributeDescription])

The policy attributes.

PolicyTypeDescription

data PolicyTypeDescription

Information about a policy type.

See: newPolicyTypeDescription smart constructor.

Constructors

PolicyTypeDescription'

Fields

policyAttributeTypeDescriptions :: Maybe [PolicyAttributeTypeDescription]

The description of the policy attributes associated with the policies defined by Elastic Load Balancing.

policyTypeName :: Maybe Text

The name of the policy type.

description :: Maybe Text

A description of the policy type.

newPolicyTypeDescription :: PolicyTypeDescription

Create a value of PolicyTypeDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policyAttributeTypeDescriptions:PolicyTypeDescription', policyTypeDescription_policyAttributeTypeDescriptions - The description of the policy attributes associated with the policies defined by Elastic Load Balancing.

$sel:policyTypeName:PolicyTypeDescription', policyTypeDescription_policyTypeName - The name of the policy type.

$sel:description:PolicyTypeDescription', policyTypeDescription_description - A description of the policy type.

policyTypeDescription_policyAttributeTypeDescriptions :: Lens' PolicyTypeDescription (Maybe [PolicyAttributeTypeDescription])

The description of the policy attributes associated with the policies defined by Elastic Load Balancing.

policyTypeDescription_policyTypeName :: Lens' PolicyTypeDescription (Maybe Text)

The name of the policy type.

policyTypeDescription_description :: Lens' PolicyTypeDescription (Maybe Text)

A description of the policy type.

SourceSecurityGroup

data SourceSecurityGroup

Information about a source security group.

See: newSourceSecurityGroup smart constructor.

Constructors

SourceSecurityGroup'

Fields

ownerAlias :: Maybe Text

The owner of the security group.

groupName :: Maybe Text

The name of the security group.

newSourceSecurityGroup :: SourceSecurityGroup

Create a value of SourceSecurityGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ownerAlias:SourceSecurityGroup', sourceSecurityGroup_ownerAlias - The owner of the security group.

$sel:groupName:SourceSecurityGroup', sourceSecurityGroup_groupName - The name of the security group.

sourceSecurityGroup_ownerAlias :: Lens' SourceSecurityGroup (Maybe Text)

The owner of the security group.

sourceSecurityGroup_groupName :: Lens' SourceSecurityGroup (Maybe Text)

The name of the security group.

Tag

data Tag

Information about a tag.

See: newTag smart constructor.

Constructors

Tag'

Fields

value :: Maybe Text

The value of the tag.

key :: Text

The key of the tag.

newTag

Arguments

:: Text
$sel:key:Tag'

-> Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:Tag', tag_value - The value of the tag.

$sel:key:Tag', tag_key - The key of the tag.

tag_value :: Lens' Tag (Maybe Text)

The value of the tag.

tag_key :: Lens' Tag Text

The key of the tag.

TagDescription

data TagDescription

The tags associated with a load balancer.

See: newTagDescription smart constructor.

Constructors

TagDescription'

Fields

tags :: Maybe (NonEmpty Tag)

The tags.

loadBalancerName :: Maybe Text

The name of the load balancer.

newTagDescription :: TagDescription

Create a value of TagDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:TagDescription', tagDescription_tags - The tags.

$sel:loadBalancerName:TagDescription', tagDescription_loadBalancerName - The name of the load balancer.

tagDescription_tags :: Lens' TagDescription (Maybe (NonEmpty Tag))

The tags.

tagDescription_loadBalancerName :: Lens' TagDescription (Maybe Text)

The name of the load balancer.

TagKeyOnly

data TagKeyOnly

The key of a tag.

See: newTagKeyOnly smart constructor.

Constructors

TagKeyOnly'

Fields

key :: Maybe Text

The name of the key.

newTagKeyOnly :: TagKeyOnly

Create a value of TagKeyOnly with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:TagKeyOnly', tagKeyOnly_key - The name of the key.

tagKeyOnly_key :: Lens' TagKeyOnly (Maybe Text)

The name of the key.