Network.AWS.ELB.Types.LoadBalancerDescription

Description

Documentation

data LoadBalancerDescription

Information about a load balancer.

See: newLoadBalancerDescription smart constructor.

Constructors

LoadBalancerDescription'

Fields

canonicalHostedZoneNameID :: Maybe Text

The ID of the Amazon Route 53 hosted zone for the load balancer.

availabilityZones :: Maybe [Text]

The Availability Zones for the load balancer.

backendServerDescriptions :: Maybe [BackendServerDescription]

Information about your EC2 instances.

policies :: Maybe Policies

The policies defined for the load balancer.

scheme :: Maybe Text

The type of load balancer. Valid only for load balancers in a VPC.

If Scheme is internet-facing, the load balancer has a public DNS name that resolves to a public IP address.

If Scheme is internal, the load balancer has a public DNS name that resolves to a private IP address.

createdTime :: Maybe ISO8601

The date and time the load balancer was created.

instances :: Maybe [Instance]

The IDs of the instances for the load balancer.

securityGroups :: Maybe [Text]

The security groups for the load balancer. Valid only for load balancers in a VPC.

sourceSecurityGroup :: Maybe SourceSecurityGroup

The security group for the load balancer, which you can use as part of your inbound rules for your registered instances. To only allow traffic from load balancers, add a security group rule that specifies this source security group as the inbound source.

dNSName :: Maybe Text

The DNS name of the load balancer.

listenerDescriptions :: Maybe [ListenerDescription]

The listeners for the load balancer.

subnets :: Maybe [Text]

The IDs of the subnets for the load balancer.

vPCId :: Maybe Text

The ID of the VPC for the load balancer.

loadBalancerName :: Maybe Text

The name of the load balancer.

healthCheck :: Maybe HealthCheck

Information about the health checks conducted on the load balancer.

canonicalHostedZoneName :: Maybe Text

The DNS name of the load balancer.

For more information, see Configure a Custom Domain Name in the Classic Load Balancers Guide.

newLoadBalancerDescription :: LoadBalancerDescription

Create a value of LoadBalancerDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:canonicalHostedZoneNameID:LoadBalancerDescription', loadBalancerDescription_canonicalHostedZoneNameID - The ID of the Amazon Route 53 hosted zone for the load balancer.

$sel:availabilityZones:LoadBalancerDescription', loadBalancerDescription_availabilityZones - The Availability Zones for the load balancer.

$sel:backendServerDescriptions:LoadBalancerDescription', loadBalancerDescription_backendServerDescriptions - Information about your EC2 instances.

$sel:policies:LoadBalancerDescription', loadBalancerDescription_policies - The policies defined for the load balancer.

$sel:scheme:LoadBalancerDescription', loadBalancerDescription_scheme - The type of load balancer. Valid only for load balancers in a VPC.

If Scheme is internet-facing, the load balancer has a public DNS name that resolves to a public IP address.

If Scheme is internal, the load balancer has a public DNS name that resolves to a private IP address.

$sel:createdTime:LoadBalancerDescription', loadBalancerDescription_createdTime - The date and time the load balancer was created.

$sel:instances:LoadBalancerDescription', loadBalancerDescription_instances - The IDs of the instances for the load balancer.

$sel:securityGroups:LoadBalancerDescription', loadBalancerDescription_securityGroups - The security groups for the load balancer. Valid only for load balancers in a VPC.

$sel:sourceSecurityGroup:LoadBalancerDescription', loadBalancerDescription_sourceSecurityGroup - The security group for the load balancer, which you can use as part of your inbound rules for your registered instances. To only allow traffic from load balancers, add a security group rule that specifies this source security group as the inbound source.

$sel:dNSName:LoadBalancerDescription', loadBalancerDescription_dNSName - The DNS name of the load balancer.

$sel:listenerDescriptions:LoadBalancerDescription', loadBalancerDescription_listenerDescriptions - The listeners for the load balancer.

$sel:subnets:LoadBalancerDescription', loadBalancerDescription_subnets - The IDs of the subnets for the load balancer.

$sel:vPCId:LoadBalancerDescription', loadBalancerDescription_vPCId - The ID of the VPC for the load balancer.

$sel:loadBalancerName:LoadBalancerDescription', loadBalancerDescription_loadBalancerName - The name of the load balancer.

$sel:healthCheck:LoadBalancerDescription', loadBalancerDescription_healthCheck - Information about the health checks conducted on the load balancer.

$sel:canonicalHostedZoneName:LoadBalancerDescription', loadBalancerDescription_canonicalHostedZoneName - The DNS name of the load balancer.

For more information, see Configure a Custom Domain Name in the Classic Load Balancers Guide.

loadBalancerDescription_canonicalHostedZoneNameID :: Lens' LoadBalancerDescription (Maybe Text)

The ID of the Amazon Route 53 hosted zone for the load balancer.

loadBalancerDescription_availabilityZones :: Lens' LoadBalancerDescription (Maybe [Text])

The Availability Zones for the load balancer.

loadBalancerDescription_backendServerDescriptions :: Lens' LoadBalancerDescription (Maybe [BackendServerDescription])

Information about your EC2 instances.

loadBalancerDescription_policies :: Lens' LoadBalancerDescription (Maybe Policies)

The policies defined for the load balancer.

loadBalancerDescription_scheme :: Lens' LoadBalancerDescription (Maybe Text)

The type of load balancer. Valid only for load balancers in a VPC.

If Scheme is internet-facing, the load balancer has a public DNS name that resolves to a public IP address.

If Scheme is internal, the load balancer has a public DNS name that resolves to a private IP address.

loadBalancerDescription_createdTime :: Lens' LoadBalancerDescription (Maybe UTCTime)

The date and time the load balancer was created.

loadBalancerDescription_instances :: Lens' LoadBalancerDescription (Maybe [Instance])

The IDs of the instances for the load balancer.

loadBalancerDescription_securityGroups :: Lens' LoadBalancerDescription (Maybe [Text])

The security groups for the load balancer. Valid only for load balancers in a VPC.

loadBalancerDescription_sourceSecurityGroup :: Lens' LoadBalancerDescription (Maybe SourceSecurityGroup)

The security group for the load balancer, which you can use as part of your inbound rules for your registered instances. To only allow traffic from load balancers, add a security group rule that specifies this source security group as the inbound source.

loadBalancerDescription_dNSName :: Lens' LoadBalancerDescription (Maybe Text)

The DNS name of the load balancer.

loadBalancerDescription_listenerDescriptions :: Lens' LoadBalancerDescription (Maybe [ListenerDescription])

The listeners for the load balancer.

loadBalancerDescription_subnets :: Lens' LoadBalancerDescription (Maybe [Text])

The IDs of the subnets for the load balancer.

loadBalancerDescription_vPCId :: Lens' LoadBalancerDescription (Maybe Text)

The ID of the VPC for the load balancer.

loadBalancerDescription_loadBalancerName :: Lens' LoadBalancerDescription (Maybe Text)

The name of the load balancer.

loadBalancerDescription_healthCheck :: Lens' LoadBalancerDescription (Maybe HealthCheck)

Information about the health checks conducted on the load balancer.

loadBalancerDescription_canonicalHostedZoneName :: Lens' LoadBalancerDescription (Maybe Text)

The DNS name of the load balancer.

For more information, see Configure a Custom Domain Name in the Classic Load Balancers Guide.