Network.AWS.ELB.V2012_06_01

Description

Elastic Load Balancing automatically distributes incoming application traffic across multiple Amazon EC2 instances. It enables you to achieve greater levels of fault tolerance in your applications, seamlessly providing the required amount of load balancing capacity needed to distribute application traffic.

Documentation

data ApplySecurityGroupsToLoadBalancer

Constructors

ApplySecurityGroupsToLoadBalancer

Fields

_asgtlbiLoadBalancerName :: Text

The name associated with the load balancer. The name must be unique within the set of load balancers associated with your AWS account.

_asgtlbiSecurityGroups :: [Text]

A list of security group IDs to associate with your load balancer in VPC. The security group IDs must be provided as the ID and not the security group name (For example, sg-1234).

asgtlbiSecurityGroups :: Lens' ApplySecurityGroupsToLoadBalancer [Text]

asgtlbiLoadBalancerName :: Lens' ApplySecurityGroupsToLoadBalancer Text

data ApplySecurityGroupsToLoadBalancerResponse

Constructors

ApplySecurityGroupsToLoadBalancerResponse

Fields

_asgtlboSecurityGroups :: [Text]

A list of security group IDs associated with your load balancer.

asgtlboSecurityGroups :: Iso' ApplySecurityGroupsToLoadBalancerResponse [Text]

data AttachLoadBalancerToSubnets

Constructors

AttachLoadBalancerToSubnets

Fields

_albtsiLoadBalancerName :: Text

The name associated with the load balancer. The name must be unique within the set of load balancers associated with your AWS account.

_albtsiSubnets :: [Text]

A list of subnet IDs to add for the load balancer. You can add only one subnet per Availability Zone.

albtsiSubnets :: Lens' AttachLoadBalancerToSubnets [Text]

albtsiLoadBalancerName :: Lens' AttachLoadBalancerToSubnets Text

data AttachLoadBalancerToSubnetsResponse

Constructors

AttachLoadBalancerToSubnetsResponse

Fields

_albtsoSubnets :: [Text]

A list of subnet IDs attached to the load balancer.

albtsoSubnets :: Iso' AttachLoadBalancerToSubnetsResponse [Text]

data ConfigureHealthCheck

Constructors

ConfigureHealthCheck

Fields

_chciLoadBalancerName :: Text

The mnemonic name associated with the load balancer. The name must be unique within the set of load balancers associated with your AWS account.

_chciHealthCheck :: HealthCheck

A structure containing the configuration information for the new healthcheck.

chciLoadBalancerName :: Lens' ConfigureHealthCheck Text

chciHealthCheck :: Lens' ConfigureHealthCheck HealthCheck

data ConfigureHealthCheckResponse

Constructors

ConfigureHealthCheckResponse

Fields

_chcoHealthCheck :: Maybe HealthCheck

The updated healthcheck for the instances.

chcoHealthCheck :: Iso' ConfigureHealthCheckResponse (Maybe HealthCheck)

data CreateAppCookieStickinessPolicy

Constructors

CreateAppCookieStickinessPolicy

Fields

_cacspiLoadBalancerName :: Text

The name of the load balancer.

_cacspiCookieName :: Text

Name of the application cookie used for stickiness.

_cacspiPolicyName :: Text

The name of the policy being created. The name must be unique within the set of policies for this load balancer.

cacspiPolicyName :: Lens' CreateAppCookieStickinessPolicy Text

cacspiLoadBalancerName :: Lens' CreateAppCookieStickinessPolicy Text

cacspiCookieName :: Lens' CreateAppCookieStickinessPolicy Text

data CreateAppCookieStickinessPolicyResponse

Constructors

CreateAppCookieStickinessPolicyResponse

createLBCookieStickinessPolicy

Arguments

:: Text
_clbcspiLoadBalancerName

-> Text
_clbcspiPolicyName

-> CreateLBCookieStickinessPolicy

Minimum specification for a CreateLBCookieStickinessPolicy request.

data CreateLBCookieStickinessPolicy

Constructors

CreateLBCookieStickinessPolicy

Fields

_clbcspiLoadBalancerName :: Text

The name associated with the load balancer.

_clbcspiPolicyName :: Text

The name of the policy being created. The name must be unique within the set of policies for this load balancer.

_clbcspiCookieExpirationPeriod :: Maybe Integer

The time period in seconds after which the cookie should be considered stale. Not specifying this parameter indicates that the sticky session will last for the duration of the browser session.

clbcspiPolicyName :: Lens' CreateLBCookieStickinessPolicy Text

clbcspiLoadBalancerName :: Lens' CreateLBCookieStickinessPolicy Text

clbcspiCookieExpirationPeriod :: Lens' CreateLBCookieStickinessPolicy (Maybe Integer)

data CreateLBCookieStickinessPolicyResponse

Constructors

CreateLBCookieStickinessPolicyResponse

createLoadBalancer

Arguments

:: Text
_capiLoadBalancerName

-> [Listener]
_capiListeners

-> CreateLoadBalancer

Minimum specification for a CreateLoadBalancer request.

data CreateLoadBalancer

Constructors

CreateLoadBalancer

Fields

_capiLoadBalancerName :: Text

The name associated with the load balancer. The name must be unique within your set of load balancers.

_capiListeners :: [Listener]

A list of the following tuples: Protocol, LoadBalancerPort, InstanceProtocol, InstancePort, and SSLCertificateId.

_capiAvailabilityZones :: [Text]

A list of Availability Zones. At least one Availability Zone must be specified. Specified Availability Zones must be in the same EC2 Region as the load balancer. Traffic will be equally distributed across all zones. You can later add more Availability Zones after the creation of the load balancer by calling EnableAvailabilityZonesForLoadBalancer action.

_capiScheme :: Maybe Text

The type of a load balancer. By default, Elastic Load Balancing creates an Internet-facing load balancer with a publicly resolvable DNS name, which resolves to public IP addresses. For more informationabout Internet-facing and Internal load balancers, see Internet-facing and Internal Load Balancers. Specify the value internal for this option to create an internal load balancer with a DNS name that resolves to private IP addresses. This option is only available for load balancers created within EC2-VPC.

_capiSecurityGroups :: [Text]

The security groups to assign to your load balancer within your VPC.

_capiSubnets :: [Text]

A list of subnet IDs in your VPC to attach to your load balancer. Specify one subnet per Availability Zone.

capiSubnets :: Lens' CreateLoadBalancer [Text]

capiSecurityGroups :: Lens' CreateLoadBalancer [Text]

capiScheme :: Lens' CreateLoadBalancer (Maybe Text)

capiLoadBalancerName :: Lens' CreateLoadBalancer Text

capiListeners :: Lens' CreateLoadBalancer [Listener]

capiAvailabilityZones :: Lens' CreateLoadBalancer [Text]

data CreateLoadBalancerResponse

Constructors

CreateLoadBalancerResponse

Fields

_capoDNSName :: Maybe Text

The DNS name for the load balancer.

capoDNSName :: Iso' CreateLoadBalancerResponse (Maybe Text)

data CreateLoadBalancerListeners

Constructors

CreateLoadBalancerListeners

Fields

_clbliLoadBalancerName :: Text

The name of the load balancer.

_clbliListeners :: [Listener]

A list of LoadBalancerPort, InstancePort, Protocol, InstanceProtocol, and SSLCertificateId items.

clbliLoadBalancerName :: Lens' CreateLoadBalancerListeners Text

clbliListeners :: Lens' CreateLoadBalancerListeners [Listener]

data CreateLoadBalancerListenersResponse

Constructors

CreateLoadBalancerListenersResponse

createLoadBalancerPolicy

Arguments

:: Text
_clbpiLoadBalancerName

-> Text
_clbpiPolicyName

-> Text
_clbpiPolicyTypeName

-> CreateLoadBalancerPolicy

Minimum specification for a CreateLoadBalancerPolicy request.

data CreateLoadBalancerPolicy

Constructors

CreateLoadBalancerPolicy

Fields

_clbpiLoadBalancerName :: Text

The name associated with the LoadBalancer for which the policy is being created.

_clbpiPolicyName :: Text

The name of the load balancer policy being created. The name must be unique within the set of policies for this load balancer.

_clbpiPolicyTypeName :: Text

The name of the base policy type being used to create this policy. To get the list of policy types, use the DescribeLoadBalancerPolicyTypes action.

_clbpiPolicyAttributes :: [PolicyAttribute]

A list of attributes associated with the policy being created.

clbpiPolicyTypeName :: Lens' CreateLoadBalancerPolicy Text

clbpiPolicyName :: Lens' CreateLoadBalancerPolicy Text

clbpiPolicyAttributes :: Lens' CreateLoadBalancerPolicy [PolicyAttribute]

clbpiLoadBalancerName :: Lens' CreateLoadBalancerPolicy Text

data CreateLoadBalancerPolicyResponse

Constructors

CreateLoadBalancerPolicyResponse

data DeleteLoadBalancer

Constructors

DeleteLoadBalancer

Fields

_dapjLoadBalancerName :: Text

The name associated with the load balancer.

dapjLoadBalancerName :: Iso' DeleteLoadBalancer Text

data DeleteLoadBalancerResponse

Constructors

DeleteLoadBalancerResponse

data DeleteLoadBalancerListeners

Constructors

DeleteLoadBalancerListeners

Fields

_dlbliLoadBalancerName :: Text

The mnemonic name associated with the load balancer.

_dlbliLoadBalancerPorts :: [Integer]

The client port number(s) of the load balancer listener(s) to be removed.

dlbliLoadBalancerPorts :: Lens' DeleteLoadBalancerListeners [Integer]

dlbliLoadBalancerName :: Lens' DeleteLoadBalancerListeners Text

data DeleteLoadBalancerListenersResponse

Constructors

DeleteLoadBalancerListenersResponse

data DeleteLoadBalancerPolicy

Constructors

DeleteLoadBalancerPolicy

Fields

_dlbpjLoadBalancerName :: Text

The mnemonic name associated with the load balancer.

_dlbpjPolicyName :: Text

The mnemonic name for the policy being deleted.

dlbpjPolicyName :: Lens' DeleteLoadBalancerPolicy Text

dlbpjLoadBalancerName :: Lens' DeleteLoadBalancerPolicy Text

data DeleteLoadBalancerPolicyResponse

Constructors

DeleteLoadBalancerPolicyResponse

data DeregisterInstancesFromLoadBalancer

Constructors

DeregisterInstancesFromLoadBalancer

Fields

_depiLoadBalancerName :: Text

The name associated with the load balancer.

_depiInstances :: [Instance]

A list of EC2 instance IDs consisting of all instances to be deregistered.

depiLoadBalancerName :: Lens' DeregisterInstancesFromLoadBalancer Text

depiInstances :: Lens' DeregisterInstancesFromLoadBalancer [Instance]

data DeregisterInstancesFromLoadBalancerResponse

Constructors

DeregisterInstancesFromLoadBalancerResponse

Fields

_depoInstances :: [Instance]

An updated list of remaining instances registered with the load balancer.

depoInstances :: Iso' DeregisterInstancesFromLoadBalancerResponse [Instance]

describeInstanceHealth

Arguments

:: Text
_depsiLoadBalancerName

-> DescribeInstanceHealth

Minimum specification for a DescribeInstanceHealth request.

data DescribeInstanceHealth

Constructors

DescribeInstanceHealth

Fields

_depsiLoadBalancerName :: Text

The name of the load balancer.

_depsiInstances :: [Instance]

A list of instance IDs whose states are being queried.

depsiLoadBalancerName :: Lens' DescribeInstanceHealth Text

depsiInstances :: Lens' DescribeInstanceHealth [Instance]

data DescribeInstanceHealthResponse

Constructors

DescribeInstanceHealthResponse

Fields

_depsoInstanceStates :: [InstanceState]

A list containing health information for the specified instances.

depsoInstanceStates :: Iso' DescribeInstanceHealthResponse [InstanceState]

data DescribeLoadBalancerAttributes

Constructors

DescribeLoadBalancerAttributes

Fields

_dlbaiLoadBalancerName :: Text

The name of the load balancer.

dlbaiLoadBalancerName :: Iso' DescribeLoadBalancerAttributes Text

data DescribeLoadBalancerAttributesResponse

Constructors

DescribeLoadBalancerAttributesResponse

Fields

_dlbaoLoadBalancerAttributes :: Maybe LoadBalancerAttributes

The load balancer attributes structure.

dlbaoLoadBalancerAttributes :: Iso' DescribeLoadBalancerAttributesResponse (Maybe LoadBalancerAttributes)

describeLoadBalancerPolicies :: DescribeLoadBalancerPolicies

Minimum specification for a DescribeLoadBalancerPolicies request.

data DescribeLoadBalancerPolicies

Constructors

DescribeLoadBalancerPolicies

Fields

_dlbpiLoadBalancerName :: Maybe Text

The mnemonic name associated with the load balancer. If no name is specified, the operation returns the attributes of either all the sample policies pre-defined by Elastic Load Balancing or the specified sample polices.

_dlbpiPolicyNames :: [Text]

The names of load balancer policies you've created or Elastic Load Balancing sample policy names.

dlbpiPolicyNames :: Lens' DescribeLoadBalancerPolicies [Text]

dlbpiLoadBalancerName :: Lens' DescribeLoadBalancerPolicies (Maybe Text)

data DescribeLoadBalancerPoliciesResponse

Constructors

DescribeLoadBalancerPoliciesResponse

Fields

_dlbpoPolicyDescriptions :: [PolicyDescription]

A list of policy description structures.

dlbpoPolicyDescriptions :: Iso' DescribeLoadBalancerPoliciesResponse [PolicyDescription]

describeLoadBalancerPolicyTypes :: DescribeLoadBalancerPolicyTypes

Minimum specification for a DescribeLoadBalancerPolicyTypes request.

data DescribeLoadBalancerPolicyTypes

Constructors

DescribeLoadBalancerPolicyTypes

Fields

_dlbptiPolicyTypeNames :: [Text]

Specifies the name of the policy types. If no names are specified, returns the description of all the policy types defined by Elastic Load Balancing service.

dlbptiPolicyTypeNames :: Iso' DescribeLoadBalancerPolicyTypes [Text]

data DescribeLoadBalancerPolicyTypesResponse

Constructors

DescribeLoadBalancerPolicyTypesResponse

Fields

_dlbptoPolicyTypeDescriptions :: [PolicyTypeDescription]

List of policy type description structures of the specified policy type. If no policy type names are specified, returns the description of all the policy types defined by Elastic Load Balancing service.

dlbptoPolicyTypeDescriptions :: Iso' DescribeLoadBalancerPolicyTypesResponse [PolicyTypeDescription]

describeLoadBalancers :: DescribeLoadBalancers

Minimum specification for a DescribeLoadBalancers request.

data DescribeLoadBalancers

Constructors

DescribeLoadBalancers

Fields

_dapiLoadBalancerNames :: [Text]

A list of load balancer names associated with the account.

_dapiMarker :: Maybe Text

An optional parameter reserved for future use.

_dapiPageSize :: Maybe Integer

The size of the page returned by this call.

dapiPageSize :: Lens' DescribeLoadBalancers (Maybe Integer)

dapiMarker :: Lens' DescribeLoadBalancers (Maybe Text)

dapiLoadBalancerNames :: Lens' DescribeLoadBalancers [Text]

data DescribeLoadBalancersResponse

Constructors

DescribeLoadBalancersResponse

Fields

_dapoLoadBalancerDescriptions :: [LoadBalancerDescription]

A list of load balancer description structures.

_dapoNextMarker :: Maybe Text

An optional parameter reserved for future use.

dapoNextMarker :: Lens' DescribeLoadBalancersResponse (Maybe Text)

dapoLoadBalancerDescriptions :: Lens' DescribeLoadBalancersResponse [LoadBalancerDescription]

data DetachLoadBalancerFromSubnets

Constructors

DetachLoadBalancerFromSubnets

Fields

_dlbfsiLoadBalancerName :: Text

The name associated with the load balancer to be detached.

_dlbfsiSubnets :: [Text]

A list of subnet IDs to remove from the set of configured subnets for the load balancer.

dlbfsiSubnets :: Lens' DetachLoadBalancerFromSubnets [Text]

dlbfsiLoadBalancerName :: Lens' DetachLoadBalancerFromSubnets Text

data DetachLoadBalancerFromSubnetsResponse

Constructors

DetachLoadBalancerFromSubnetsResponse

Fields

_dlbfsoSubnets :: [Text]

A list of subnet IDs the load balancer is now attached to.

dlbfsoSubnets :: Iso' DetachLoadBalancerFromSubnetsResponse [Text]

data DisableAvailabilityZonesForLoadBalancer

Constructors

DisableAvailabilityZonesForLoadBalancer

Fields

_raziLoadBalancerName :: Text

The name associated with the load balancer.

_raziAvailabilityZones :: [Text]

A list of Availability Zones to be removed from the load balancer. There must be at least one Availability Zone registered with a load balancer at all times. Specified Availability Zones must be in the same region.

raziLoadBalancerName :: Lens' DisableAvailabilityZonesForLoadBalancer Text

raziAvailabilityZones :: Lens' DisableAvailabilityZonesForLoadBalancer [Text]

data DisableAvailabilityZonesForLoadBalancerResponse

Constructors

DisableAvailabilityZonesForLoadBalancerResponse

Fields

_razoAvailabilityZones :: [Text]

A list of updated Availability Zones for the load balancer.

razoAvailabilityZones :: Iso' DisableAvailabilityZonesForLoadBalancerResponse [Text]

data EnableAvailabilityZonesForLoadBalancer

Constructors

EnableAvailabilityZonesForLoadBalancer

Fields

_aaziLoadBalancerName :: Text

The name associated with the load balancer.

_aaziAvailabilityZones :: [Text]

A list of new Availability Zones for the load balancer. Each Availability Zone must be in the same region as the load balancer.

aaziLoadBalancerName :: Lens' EnableAvailabilityZonesForLoadBalancer Text

aaziAvailabilityZones :: Lens' EnableAvailabilityZonesForLoadBalancer [Text]

data EnableAvailabilityZonesForLoadBalancerResponse

Constructors

EnableAvailabilityZonesForLoadBalancerResponse

Fields

_aazoAvailabilityZones :: [Text]

An updated list of Availability Zones for the load balancer.

aazoAvailabilityZones :: Iso' EnableAvailabilityZonesForLoadBalancerResponse [Text]

data ModifyLoadBalancerAttributes

Constructors

ModifyLoadBalancerAttributes

Fields

_mlbaiLoadBalancerName :: Text

The name of the load balancer.

_mlbaiLoadBalancerAttributes :: LoadBalancerAttributes

Attributes of the load balancer.

mlbaiLoadBalancerName :: Lens' ModifyLoadBalancerAttributes Text

mlbaiLoadBalancerAttributes :: Lens' ModifyLoadBalancerAttributes LoadBalancerAttributes

data ModifyLoadBalancerAttributesResponse

Constructors

ModifyLoadBalancerAttributesResponse

Fields

_mlbaoLoadBalancerName :: Maybe Text

The name of the load balancer.

_mlbaoLoadBalancerAttributes :: Maybe LoadBalancerAttributes

The LoadBalancerAttributes data type.

mlbaoLoadBalancerName :: Lens' ModifyLoadBalancerAttributesResponse (Maybe Text)

mlbaoLoadBalancerAttributes :: Lens' ModifyLoadBalancerAttributesResponse (Maybe LoadBalancerAttributes)

data RegisterInstancesWithLoadBalancer

Constructors

RegisterInstancesWithLoadBalancer

Fields

_repiLoadBalancerName :: Text

The name associated with the load balancer. The name must be unique within your set of load balancers.

_repiInstances :: [Instance]

A list of instance IDs that should be registered with the load balancer.

repiLoadBalancerName :: Lens' RegisterInstancesWithLoadBalancer Text

repiInstances :: Lens' RegisterInstancesWithLoadBalancer [Instance]

data RegisterInstancesWithLoadBalancerResponse

Constructors

RegisterInstancesWithLoadBalancerResponse

Fields

_repoInstances :: [Instance]

An updated list of instances for the load balancer.

repoInstances :: Iso' RegisterInstancesWithLoadBalancerResponse [Instance]

data SetLoadBalancerListenerSSLCertificate

Constructors

SetLoadBalancerListenerSSLCertificate

Fields

_slblsslciLoadBalancerName :: Text

The name of the load balancer.

_slblsslciLoadBalancerPort :: Integer

The port that uses the specified SSL certificate.

_slblsslciSSLCertificateId :: Text

The Amazon Resource Number (ARN) of the SSL certificate chain to use. For more information on SSL certificates, see Managing Server Certificates in the AWS Identity and Access Management User Guide.

slblsslciSSLCertificateId :: Lens' SetLoadBalancerListenerSSLCertificate Text

slblsslciLoadBalancerPort :: Lens' SetLoadBalancerListenerSSLCertificate Integer

slblsslciLoadBalancerName :: Lens' SetLoadBalancerListenerSSLCertificate Text

data SetLoadBalancerListenerSSLCertificateResponse

Constructors

SetLoadBalancerListenerSSLCertificateResponse

data SetLoadBalancerPoliciesForBackendServer

Constructors

SetLoadBalancerPoliciesForBackendServer

Fields

_slbpfbsiLoadBalancerName :: Text

The mnemonic name associated with the load balancer. This name must be unique within the set of your load balancers.

_slbpfbsiInstancePort :: Integer

The port number associated with the back-end server.

_slbpfbsiPolicyNames :: [Text]

List of policy names to be set. If the list is empty, then all current polices are removed from the back-end server.

slbpfbsiPolicyNames :: Lens' SetLoadBalancerPoliciesForBackendServer [Text]

slbpfbsiLoadBalancerName :: Lens' SetLoadBalancerPoliciesForBackendServer Text

slbpfbsiInstancePort :: Lens' SetLoadBalancerPoliciesForBackendServer Integer

data SetLoadBalancerPoliciesForBackendServerResponse

Constructors

SetLoadBalancerPoliciesForBackendServerResponse

data SetLoadBalancerPoliciesOfListener

Constructors

SetLoadBalancerPoliciesOfListener

Fields

_slbpoliLoadBalancerName :: Text

The name of the load balancer.

_slbpoliLoadBalancerPort :: Integer

The external port of the load balancer to associate the policy.

_slbpoliPolicyNames :: [Text]

List of policies to be associated with the listener. If the list is empty, the current policy is removed from the listener.

slbpoliPolicyNames :: Lens' SetLoadBalancerPoliciesOfListener [Text]

slbpoliLoadBalancerPort :: Lens' SetLoadBalancerPoliciesOfListener Integer

slbpoliLoadBalancerName :: Lens' SetLoadBalancerPoliciesOfListener Text

data SetLoadBalancerPoliciesOfListenerResponse

Constructors

SetLoadBalancerPoliciesOfListenerResponse

data ELB

Supported version (2012-06-01) of the Elastic Load Balancing service.

xmlOptions :: Tagged a XMLOptions

newtype ConnectionSettings

The name of the load balancer attribute. By default, Elastic Load Balancing maintains a 60-second idle connection timeout for both front-end and back-end connections of your load balancer. If ConnectionSettings attribute is set, Elastic Load Balancing will allow the connections to remain idle (no data is sent over the connection) for the specified duration. For more information, see Configure Idle Connection Timeout.

Constructors

ConnectionSettings

Fields

_csIdleTimeout :: Integer

Specifies the maximum time the connection is allowed to be idle (no data has been sent over the connection) before it is closed by the load balancer.

newtype CrossZoneLoadBalancing

The name of the load balancer attribute. If enabled, the load balancer routes the request traffic evenly across all back-end instances regardless of the Availability Zones. For more information, see Enable Cross-Zone Load Balancing.

Constructors

CrossZoneLoadBalancing

Fields

_czlbEnabled :: Bool

Specifies whether cross-zone load balancing is enabled for the load balancer.

newtype Instance

The Instance data type.

Constructors

Instance

Fields

_jInstanceId :: Maybe Text

Provides an EC2 instance ID.

data AccessLog

The name of the load balancer attribute. If enabled, the load balancer captures detailed information of all the requests and delivers the information to the Amazon S3 bucket you specify. For more information, see Enable Access Logs.

Constructors

AccessLog

Fields

_alEmitInterval :: Maybe Integer

The interval for publishing the access logs. You can specify an interval of either 5 minutes or 60 minutes. Default: 60 minutes.

_alEnabled :: Bool

Specifies whether access log is enabled for the load balancer.

_alS3BucketPrefix :: Maybe Text

The logical hierarchy you created for your Amazon S3 bucket, for example my-bucket-prefix/prod. If the prefix is not provided, the log is placed at the root level of the bucket.

_alS3BucketName :: Maybe Text

The name of the Amazon S3 bucket where the access logs are stored.

data AppCookieStickinessPolicy

The AppCookieStickinessPolicy data type.

Constructors

AppCookieStickinessPolicy

Fields

_acsqPolicyName :: Maybe Text

The mnemonic name for the policy being created. The name must be unique within a set of policies for this load balancer.

_acsqCookieName :: Maybe Text

The name of the application cookie used for stickiness.

data BackendServerDescription

This data type is used as a response element in the DescribeLoadBalancers action to describe the configuration of the back-end server.

Constructors

BackendServerDescription

Fields

_bsePolicyNames :: [Text]

Provides a list of policy names enabled for the back-end server.

_bseInstancePort :: Maybe Integer

Provides the port on which the back-end server is listening.

data ConnectionDraining

The name of the load balancer attribute. If enabled, the load balancer allows existing requests to complete before the load balancer shifts traffic away from a deregistered or unhealthy back-end instance. For more information, see Enable Connection Draining.

Constructors

ConnectionDraining

Fields

_cdEnabled :: Bool

Specifies whether connection draining is enabled for the load balancer.

_cdTimeout :: Maybe Integer

Specifies the maximum time (in seconds) to keep the existing connections open before deregistering the instances.

data HealthCheck

Specifies information regarding the various health probes conducted on the load balancer.

Constructors

HealthCheck

Fields

_hcHealthyThreshold :: Integer

Specifies the number of consecutive health probe successes required before moving the instance to the Healthy state.

_hcInterval :: Integer

Specifies the approximate interval, in seconds, between health checks of an individual instance.

_hcTimeout :: Integer

Specifies the amount of time, in seconds, during which no response means a failed health probe. This value must be less than the Interval value.

_hcUnhealthyThreshold :: Integer

Specifies the number of consecutive health probe failures required before moving the instance to the Unhealthy state.

_hcTarget :: Text

Specifies the instance being checked. The protocol is either TCP, HTTP, HTTPS, or SSL. The range of valid ports is one (1) through 65535. TCP is the default, specified as a TCP: port pair, for example TCP:5000. In this case a healthcheck simply attempts to open a TCP connection to the instance on the specified port. Failure to connect within the configured timeout is considered unhealthy. SSL is also specified as SSL: port pair, for example, SSL:5000. For HTTP or HTTPS protocol, the situation is different. You have to include a ping path in the string. HTTP is specified as a HTTP:port;/;PathToPing; grouping, for example "HTTP:80weatheruswaseattle". In this case, a HTTP GET request is issued to the instance on the given port and path. Any answer other than "200 OK" within the timeout period is considered unhealthy. The total length of the HTTP ping target needs to be 1024 16-bit Unicode characters or less.

data InstanceState

The InstanceState data type.

Constructors

InstanceState

Fields

_ivInstanceId :: Maybe Text

Provides an EC2 instance ID.

_ivState :: Maybe Text

Specifies the current state of the instance. Valid value: InService|OutOfService|Unknown.

_ivReasonCode :: Maybe Text

Provides information about the cause of OutOfService instances. Specifically, it indicates whether the cause is Elastic Load Balancing or the instance behind the load balancer. Valid value: ELB|Instance|N/A.

_ivDescription :: Maybe Text

Provides a description of the instance state.

data LBCookieStickinessPolicy

The LBCookieStickinessPolicy data type.

Constructors

LBCookieStickinessPolicy

Fields

_lbcsqPolicyName :: Maybe Text

The name for the policy being created. The name must be unique within the set of policies for this load balancer.

_lbcsqCookieExpirationPeriod :: Maybe Integer

The time period in seconds after which the cookie should be considered stale. Not specifying this parameter indicates that the stickiness session will last for the duration of the browser session.

data Listener

The Listener data type.

Constructors

Listener

Fields

_lInstanceProtocol :: Maybe Text

Specifies the protocol to use for routing traffic to back-end instances - HTTP, HTTPS, TCP, or SSL. This property cannot be modified for the life of the load balancer. If the front-end protocol is HTTP or HTTPS, InstanceProtocol has to be at the same protocol layer, i.e., HTTP or HTTPS. Likewise, if the front-end protocol is TCP or SSL, InstanceProtocol has to be TCP or SSL. If there is another listener with the same InstancePort whose InstanceProtocol is secure, i.e., HTTPS or SSL, the listener's InstanceProtocol has to be secure, i.e., HTTPS or SSL. If there is another listener with the same InstancePort whose InstanceProtocol is HTTP or TCP, the listener's InstanceProtocol must be either HTTP or TCP.

_lInstancePort :: Integer

Specifies the TCP port on which the instance server is listening. This property cannot be modified for the life of the load balancer.

_lLoadBalancerPort :: Integer

Specifies the external load balancer port number. This property cannot be modified for the life of the load balancer.

_lProtocol :: Text

Specifies the load balancer transport protocol to use for routing - HTTP, HTTPS, TCP or SSL. This property cannot be modified for the life of the load balancer.

_lSSLCertificateId :: Maybe Text

The ARN string of the server certificate. To get the ARN of the server certificate, call the AWS Identity and Access Management UploadServerCertificate API.

data ListenerDescription

The ListenerDescription data type.

Constructors

ListenerDescription

Fields

_lePolicyNames :: [Text]

A list of policies enabled for this listener. An empty list indicates that no policies are enabled.

_leListener :: Maybe Listener

The Listener data type.

data LoadBalancerAttributes

Attributes of the load balancer.

Constructors

LoadBalancerAttributes

Fields

_lbaCrossZoneLoadBalancing :: Maybe CrossZoneLoadBalancing

The name of the load balancer attribute. If enabled, the load balancer routes the request traffic evenly across all back-end instances regardless of the Availability Zones. For more information, see Enable Cross-Zone Load Balancing.

_lbaAccessLog :: Maybe AccessLog

The name of the load balancer attribute. If enabled, the load balancer captures detailed information of all the requests and delivers the information to the Amazon S3 bucket you specify. For more information, see Enable Access Logs.

_lbaConnectionSettings :: Maybe ConnectionSettings

The name of the load balancer attribute. By default, Elastic Load Balancing maintains a 60-second idle connection timeout for both front-end and back-end connections of your load balancer. If ConnectionSettings attribute is set, Elastic Load Balancing will allow the connections to remain idle (no data is sent over the connection) for the specified duration. For more information, see Configure Idle Connection Timeout.

_lbaConnectionDraining :: Maybe ConnectionDraining

The name of the load balancer attribute. If enabled, the load balancer allows existing requests to complete before the load balancer shifts traffic away from a deregistered or unhealthy back-end instance. For more information, see Enable Connection Draining.

data LoadBalancerDescription

Contains the result of a successful invocation of DescribeLoadBalancers.

Constructors

LoadBalancerDescription

Fields

_lbeSourceSecurityGroup :: Maybe SourceSecurityGroup

The security group that you can use as part of your inbound rules for your load balancer's back-end Amazon EC2 application instances. To only allow traffic from load balancers, add a security group rule to your back end instance that specifies this source security group as the inbound source.

_lbeCanonicalHostedZoneName :: Maybe Text

Provides the name of the Amazon Route 53 hosted zone that is associated with the load balancer. For information on how to associate your load balancer with a hosted zone, go to Using Domain Names With Elastic Load Balancing in the Elastic Load Balancing Developer Guide.

_lbeSecurityGroups :: [Text]

The security groups the load balancer is a member of (VPC only).

_lbeHealthCheck :: Maybe HealthCheck

Specifies information regarding the various health probes conducted on the load balancer.

_lbeLoadBalancerName :: Maybe Text

Specifies the name associated with the load balancer.

_lbeCreatedTime :: Maybe ISO8601

Provides the date and time the load balancer was created.

_lbeVPCId :: Maybe Text

Provides the ID of the VPC attached to the load balancer.

_lbeSubnets :: [Text]

Provides a list of VPC subnet IDs for the load balancer.

_lbeAvailabilityZones :: [Text]

Specifies a list of Availability Zones.

_lbeBackendServerDescriptions :: [BackendServerDescription]

Contains a list of back-end server descriptions.

_lbeCanonicalHostedZoneNameID :: Maybe Text

Provides the ID of the Amazon Route 53 hosted zone name that is associated with the load balancer. For information on how to associate or disassociate your load balancer with a hosted zone, go to Using Domain Names With Elastic Load Balancing in the Elastic Load Balancing Developer Guide.

_lbeInstances :: [Instance]

Provides a list of EC2 instance IDs for the load balancer.

_lbeScheme :: Maybe Text

Specifies the type of load balancer. If the Scheme is internet-facing, the load balancer has a publicly resolvable DNS name that resolves to public IP addresses. If the Scheme is internal, the load balancer has a publicly resolvable DNS name that resolves to private IP addresses. This option is only available for load balancers attached to an Amazon VPC.

_lbeListenerDescriptions :: [ListenerDescription]

LoadBalancerPort, InstancePort, Protocol, InstanceProtocol, and PolicyNames are returned in a list of tuples in the ListenerDescriptions element.

_lbeDNSName :: Maybe Text

Specifies the external DNS name associated with the load balancer.

_lbePolicies :: Maybe Policies

Provides a list of policies defined for the load balancer.

data Policies

Provides a list of policies defined for the load balancer.

Constructors

Policies

Fields

_rOtherPolicies :: [Text]

A list of policy names other than the stickiness policies.

_rLBCookieStickinessPolicies :: [LBCookieStickinessPolicy]

A list of LBCookieStickinessPolicy objects created with CreateAppCookieStickinessPolicy.

_rAppCookieStickinessPolicies :: [AppCookieStickinessPolicy]

A list of the AppCookieStickinessPolicy objects created with CreateAppCookieStickinessPolicy.

data PolicyAttribute

The PolicyAttribute data type. This data type contains a key/value pair that defines properties of a specific policy.

Constructors

PolicyAttribute

Fields

_pbAttributeValue :: Maybe Text

The value of the attribute associated with the policy.

_pbAttributeName :: Maybe Text

The name of the attribute associated with the policy.

data PolicyAttributeDescription

The PolicyAttributeDescription data type. This data type is used to describe the attributes and values associated with a policy.

Constructors

PolicyAttributeDescription

Fields

_paeAttributeValue :: Maybe Text

The value of the attribute associated with the policy.

_paeAttributeName :: Maybe Text

The name of the attribute associated with the policy.

data PolicyAttributeTypeDescription

The PolicyAttributeTypeDescription data type. This data type is used to describe values that are acceptable for the policy attribute.

Constructors

PolicyAttributeTypeDescription

Fields

_pateAttributeType :: Maybe Text

The type of attribute. For example, Boolean, Integer, etc.

_pateCardinality :: Maybe Text

The cardinality of the attribute. Valid Values: ONE(1) : Single value required ZERO_OR_ONE(0..1) : Up to one value can be supplied ZERO_OR_MORE(0..*) : Optional. Multiple values are allowed ONE_OR_MORE(1..*0) : Required. Multiple values are allowed.

_pateDefaultValue :: Maybe Text

The default value of the attribute, if applicable.

_pateAttributeName :: Maybe Text

The name of the attribute associated with the policy type.

_pateDescription :: Maybe Text

A human-readable description of the attribute.

data PolicyDescription

The PolicyDescription data type.

Constructors

PolicyDescription

Fields

_pePolicyName :: Maybe Text

The name of the policy associated with the load balancer.

_pePolicyAttributeDescriptions :: [PolicyAttributeDescription]

A list of policy attribute description structures.

_pePolicyTypeName :: Maybe Text

The name of the policy type associated with the load balancer.

data PolicyTypeDescription

The PolicyTypeDescription data type.

Constructors

PolicyTypeDescription

Fields

_ptePolicyTypeName :: Maybe Text

The name of the policy type.

_pteDescription :: Maybe Text

A human-readable description of the policy type.

_ptePolicyAttributeTypeDescriptions :: [PolicyAttributeTypeDescription]

The description of the policy attributes associated with the load balancer policies defined by the Elastic Load Balancing service.

data SourceSecurityGroup

The security group that you can use as part of your inbound rules for your load balancer's back-end Amazon EC2 application instances. To only allow traffic from load balancers, add a security group rule to your back end instance that specifies this source security group as the inbound source.

Constructors

SourceSecurityGroup

Fields

_ssgOwnerAlias :: Maybe Text

Owner of the source security group. Use this value for the --source-group-user parameter of the ec2-authorize command in the Amazon EC2 command line tool.

_ssgGroupName :: Maybe Text

Name of the source security group. Use this value for the --source-group parameter of the ec2-authorize command in the Amazon EC2 command line tool.

csIdleTimeout :: Iso' ConnectionSettings Integer

connectionSettings :: Iso' Integer ConnectionSettings

czlbEnabled :: Iso' CrossZoneLoadBalancing Bool

crossZoneLoadBalancing :: Iso' Bool CrossZoneLoadBalancing

jInstanceId :: Iso' Instance (Maybe Text)

instance :: Iso' (Maybe Text) Instance

alS3BucketPrefix :: Lens' AccessLog (Maybe Text)

alS3BucketName :: Lens' AccessLog (Maybe Text)

alEnabled :: Lens' AccessLog Bool

alEmitInterval :: Lens' AccessLog (Maybe Integer)

acsqPolicyName :: Lens' AppCookieStickinessPolicy (Maybe Text)

acsqCookieName :: Lens' AppCookieStickinessPolicy (Maybe Text)

bsePolicyNames :: Lens' BackendServerDescription [Text]

bseInstancePort :: Lens' BackendServerDescription (Maybe Integer)

cdTimeout :: Lens' ConnectionDraining (Maybe Integer)

cdEnabled :: Lens' ConnectionDraining Bool

hcUnhealthyThreshold :: Lens' HealthCheck Integer

hcTimeout :: Lens' HealthCheck Integer

hcTarget :: Lens' HealthCheck Text

hcInterval :: Lens' HealthCheck Integer

hcHealthyThreshold :: Lens' HealthCheck Integer

ivState :: Lens' InstanceState (Maybe Text)

ivReasonCode :: Lens' InstanceState (Maybe Text)

ivInstanceId :: Lens' InstanceState (Maybe Text)

ivDescription :: Lens' InstanceState (Maybe Text)

lbcsqPolicyName :: Lens' LBCookieStickinessPolicy (Maybe Text)

lbcsqCookieExpirationPeriod :: Lens' LBCookieStickinessPolicy (Maybe Integer)

lSSLCertificateId :: Lens' Listener (Maybe Text)

lProtocol :: Lens' Listener Text

lLoadBalancerPort :: Lens' Listener Integer

lInstanceProtocol :: Lens' Listener (Maybe Text)

lInstancePort :: Lens' Listener Integer

lePolicyNames :: Lens' ListenerDescription [Text]

leListener :: Lens' ListenerDescription (Maybe Listener)

lbaCrossZoneLoadBalancing :: Lens' LoadBalancerAttributes (Maybe CrossZoneLoadBalancing)

lbaConnectionSettings :: Lens' LoadBalancerAttributes (Maybe ConnectionSettings)

lbaConnectionDraining :: Lens' LoadBalancerAttributes (Maybe ConnectionDraining)

lbaAccessLog :: Lens' LoadBalancerAttributes (Maybe AccessLog)

lbeVPCId :: Lens' LoadBalancerDescription (Maybe Text)

lbeSubnets :: Lens' LoadBalancerDescription [Text]

lbeSourceSecurityGroup :: Lens' LoadBalancerDescription (Maybe SourceSecurityGroup)

lbeSecurityGroups :: Lens' LoadBalancerDescription [Text]

lbeScheme :: Lens' LoadBalancerDescription (Maybe Text)

lbePolicies :: Lens' LoadBalancerDescription (Maybe Policies)

lbeLoadBalancerName :: Lens' LoadBalancerDescription (Maybe Text)

lbeListenerDescriptions :: Lens' LoadBalancerDescription [ListenerDescription]

lbeInstances :: Lens' LoadBalancerDescription [Instance]

lbeHealthCheck :: Lens' LoadBalancerDescription (Maybe HealthCheck)

lbeDNSName :: Lens' LoadBalancerDescription (Maybe Text)

lbeCreatedTime :: Lens' LoadBalancerDescription (Maybe ISO8601)

lbeCanonicalHostedZoneNameID :: Lens' LoadBalancerDescription (Maybe Text)

lbeCanonicalHostedZoneName :: Lens' LoadBalancerDescription (Maybe Text)

lbeBackendServerDescriptions :: Lens' LoadBalancerDescription [BackendServerDescription]

lbeAvailabilityZones :: Lens' LoadBalancerDescription [Text]

rOtherPolicies :: Lens' Policies [Text]

rLBCookieStickinessPolicies :: Lens' Policies [LBCookieStickinessPolicy]

rAppCookieStickinessPolicies :: Lens' Policies [AppCookieStickinessPolicy]

pbAttributeValue :: Lens' PolicyAttribute (Maybe Text)

pbAttributeName :: Lens' PolicyAttribute (Maybe Text)

paeAttributeValue :: Lens' PolicyAttributeDescription (Maybe Text)

paeAttributeName :: Lens' PolicyAttributeDescription (Maybe Text)

pateDescription :: Lens' PolicyAttributeTypeDescription (Maybe Text)

pateDefaultValue :: Lens' PolicyAttributeTypeDescription (Maybe Text)

pateCardinality :: Lens' PolicyAttributeTypeDescription (Maybe Text)

pateAttributeType :: Lens' PolicyAttributeTypeDescription (Maybe Text)

pateAttributeName :: Lens' PolicyAttributeTypeDescription (Maybe Text)

pePolicyTypeName :: Lens' PolicyDescription (Maybe Text)

pePolicyName :: Lens' PolicyDescription (Maybe Text)

pePolicyAttributeDescriptions :: Lens' PolicyDescription [PolicyAttributeDescription]

ptePolicyTypeName :: Lens' PolicyTypeDescription (Maybe Text)

ptePolicyAttributeTypeDescriptions :: Lens' PolicyTypeDescription [PolicyAttributeTypeDescription]

pteDescription :: Lens' PolicyTypeDescription (Maybe Text)

ssgOwnerAlias :: Lens' SourceSecurityGroup (Maybe Text)

ssgGroupName :: Lens' SourceSecurityGroup (Maybe Text)