Network.AWS.ELB.V2012_06_01.CreateAppCookieStickinessPolicy

Description

Generates a stickiness policy with sticky session lifetimes that follow that of an application-generated cookie. This policy can be associated only with HTTP/HTTPS listeners. This policy is similar to the policy created by CreateLBCookieStickinessPolicy, except that the lifetime of the special Elastic Load Balancing cookie follows the lifetime of the application-generated cookie specified in the policy configuration. The load balancer only inserts a new stickiness cookie when the application response includes a new application cookie. If the application cookie is explicitly removed or expires, the session stops being sticky until a new application cookie is issued. An application client must receive and send two cookies: the application-generated cookie and the special Elastic Load Balancing cookie named AWSELB. This is the default behavior for many common web browsers. For more information, see Enabling Application-Controlled Session Stickiness in the Elastic Load Balancing Developer Guide. https:/elasticloadbalancing.amazonaws.com?CookieName=MyAppCookie &LoadBalancerName=MyLoadBalancer &PolicyName=MyAppStickyPolicy &Version=2012-06-01 &Action=CreateAppCookieStickinessPolicy &AUTHPARAMS 99a693e9-12b8-11e3-9ad6-bf3e4EXAMPLE.

Documentation

data CreateAppCookieStickinessPolicy

Constructors

CreateAppCookieStickinessPolicy

Fields

_cacspiLoadBalancerName :: Text

The name of the load balancer.

_cacspiCookieName :: Text

Name of the application cookie used for stickiness.

_cacspiPolicyName :: Text

The name of the policy being created. The name must be unique within the set of policies for this load balancer.

cacspiPolicyName :: Lens' CreateAppCookieStickinessPolicy Text

cacspiLoadBalancerName :: Lens' CreateAppCookieStickinessPolicy Text

cacspiCookieName :: Lens' CreateAppCookieStickinessPolicy Text

data CreateAppCookieStickinessPolicyResponse

Constructors

CreateAppCookieStickinessPolicyResponse