Network.AWS.ELB.V2012_06_01.CreateLBCookieStickinessPolicy

Description

Generates a stickiness policy with sticky session lifetimes controlled by the lifetime of the browser (user-agent) or a specified expiration period. This policy can be associated only with HTTP/HTTPS listeners. When a load balancer implements this policy, the load balancer uses a special cookie to track the backend server instance for each request. When the load balancer receives a request, it first checks to see if this cookie is present in the request. If so, the load balancer sends the request to the application server specified in the cookie. If not, the load balancer sends the request to a server that is chosen based on the existing load balancing algorithm. A cookie is inserted into the response for binding subsequent requests from the same user to that server. The validity of the cookie is based on the cookie expiration time, which is specified in the policy configuration. For more information, see Enabling Duration-Based Session Stickiness in the Elastic Load Balancing Developer Guide. https:/elasticloadbalancing.amazonaws.com?CookieExpirationPeriod=60 &LoadBalancerName=MyLoadBalancer&PolicyName=MyDurationStickyPolicy &Version=2012-06-01 &Action=CreateLBCookieStickinessPolicy &AUTHPARAMS 99a693e9-12b8-11e3-9ad6-bf3e4EXAMPLE.

Documentation

createLBCookieStickinessPolicy

Arguments

:: Text
_clbcspiLoadBalancerName

-> Text
_clbcspiPolicyName

-> CreateLBCookieStickinessPolicy

Minimum specification for a CreateLBCookieStickinessPolicy request.

data CreateLBCookieStickinessPolicy

Constructors

CreateLBCookieStickinessPolicy

Fields

_clbcspiLoadBalancerName :: Text

The name associated with the load balancer.

_clbcspiPolicyName :: Text

The name of the policy being created. The name must be unique within the set of policies for this load balancer.

_clbcspiCookieExpirationPeriod :: Maybe Integer

The time period in seconds after which the cookie should be considered stale. Not specifying this parameter indicates that the sticky session will last for the duration of the browser session.

clbcspiPolicyName :: Lens' CreateLBCookieStickinessPolicy Text

clbcspiLoadBalancerName :: Lens' CreateLBCookieStickinessPolicy Text

clbcspiCookieExpirationPeriod :: Lens' CreateLBCookieStickinessPolicy (Maybe Integer)

data CreateLBCookieStickinessPolicyResponse

Constructors

CreateLBCookieStickinessPolicyResponse