Network.AWS.ELB.V2012_06_01.DeregisterInstancesFromLoadBalancer

Description

Deregisters instances from the load balancer. Once the instance is deregistered, it will stop receiving traffic from the load balancer. In order to successfully call this API, the same account credentials as those used to create the load balancer must be provided. For more information, see De-register and Register Amazon EC2 Instances in the Elastic Load Balancing Developer Guide. You can use DescribeLoadBalancers to verify if the instance is deregistered from the load balancer. Deregister instance i-e3677ad7 from MyHTTPSLoadBalancer load balancer. https:/elasticloadbalancing.amazonaws.com?Instances.member.1.InstanceId=i-e3677ad7 &LoadBalancerName=MyHTTPSLoadBalancer &Version=2012-06-01 &Action=DeregisterInstancesFromLoadBalancer &AUTHPARAMS i-6ec63d59 83c88b9d-12b7-11e3-8b82-87b12EXAMPLE.

Documentation

data DeregisterInstancesFromLoadBalancer

Constructors

DeregisterInstancesFromLoadBalancer

Fields

_depiLoadBalancerName :: Text

The name associated with the load balancer.

_depiInstances :: [Instance]

A list of EC2 instance IDs consisting of all instances to be deregistered.

depiLoadBalancerName :: Lens' DeregisterInstancesFromLoadBalancer Text

depiInstances :: Lens' DeregisterInstancesFromLoadBalancer [Instance]

data DeregisterInstancesFromLoadBalancerResponse

Constructors

DeregisterInstancesFromLoadBalancerResponse

Fields

_depoInstances :: [Instance]

An updated list of remaining instances registered with the load balancer.

depoInstances :: Iso' DeregisterInstancesFromLoadBalancerResponse [Instance]