Network.AWS.ELB.V2012_06_01.EnableAvailabilityZonesForLoadBalancer

Description

Adds one or more EC2 Availability Zones to the load balancer. The load balancer evenly distributes requests across all its registered Availability Zones that contain instances. The new EC2 Availability Zones to be added must be in the same EC2 Region as the Availability Zones for which the load balancer was created. For more information, see Expand a Load Balanced Application to an Additional Availability Zone in the Elastic Load Balancing Developer Guide. Enable Availability Zone us-east-1c for my-test-loadbalancer. https:/elasticloadbalancing.amazonaws.com?AvailabilityZones.member.1=us-east-1c &LoadBalancerName=my-test-loadbalancer &Version=2012-06-01 &Action=EnableAvailabilityZonesForLoadBalancer &AUTHPARAMS us-east-1a us-east-1c 83c88b9d-12b7-11e3-8b82-87b12EXAMPLE.

Documentation

data EnableAvailabilityZonesForLoadBalancer

Constructors

EnableAvailabilityZonesForLoadBalancer

Fields

_aaziLoadBalancerName :: Text

The name associated with the load balancer.

_aaziAvailabilityZones :: [Text]

A list of new Availability Zones for the load balancer. Each Availability Zone must be in the same region as the load balancer.

aaziLoadBalancerName :: Lens' EnableAvailabilityZonesForLoadBalancer Text

aaziAvailabilityZones :: Lens' EnableAvailabilityZonesForLoadBalancer [Text]

data EnableAvailabilityZonesForLoadBalancerResponse

Constructors

EnableAvailabilityZonesForLoadBalancerResponse

Fields

_aazoAvailabilityZones :: [Text]

An updated list of Availability Zones for the load balancer.

aazoAvailabilityZones :: Iso' EnableAvailabilityZonesForLoadBalancerResponse [Text]