Network.AWS.ELB.V2012_06_01.ModifyLoadBalancerAttributes

Description

Modifies the attributes of a specified load balancer. You can modify the load balancer attributes, such as AccessLogs, ConnectionDraining, and CrossZoneLoadBalancing by either enabling or disabling them. Or, you can modify the load balancer attribute ConnectionSettings, by specifying an idle connection timeout value for your load balancer. For more information, see the following: Cross-Zone Load Balancing Connection Draining Access Logs. Connection Settings. Enable Cross-Zone Load Balancing https:/elasticloadbalancing.amazonaws.com?LoadBalancerName=my-test-loadbalancer &LoadBalancerAttributes.CrossZoneLoadBalancing.Enabled=true &Version=2012-06-01 &Action=ModifyLoadBalancerAttributes &AUTHPARAMS my-test-loadbalancer true 83c88b9d-12b7-11e3-8b82-87b12EXAMPLE Enable Access Log https:/elasticloadbalancing.amazonaws.com?LoadBalancerName=my-test-loadbalancer &LoadBalancerAttributes.AccessLog.Enabled=true &LoadBalancerAttributes.AccessLog.S3BucketName=my-loadbalancer-logs &LoadBalancerAttributes.AccessLog.S3BucketPrefix=my-bucket-prefix/prod &LoadBalancerAttributes.AccessLog.EmitInterval=60 &Version=2012-06-01 &Action=ModifyLoadBalancerAttributes &AUTHPARAMS my-test-loadbalancer true my-loadbalancer-logs my-bucket-prefix/prod 60 83c88b9d-12b7-11e3-8b82-87b12EXAMPLE Enable Connection Draining https:/elasticloadbalancing.amazonaws.com?LoadBalancerName=my-test-loadbalancer &LoadBalancerAttributes.ConnectionDraining.Enabled=true &LoadBalancerAttributes.ConnectionDraining.Timeout=60 &Version=2012-06-01 &Action=ModifyLoadBalancerAttributes &AUTHPARAMS my-test-loadbalancer true 60 83c88b9d-12b7-11e3-8b82-87b12EXAMPLE Configure Connection Settings https:/elasticloadbalancing.amazonaws.com?LoadBalancerName=my-test-loadbalancer &LoadBalancerAttributes.ConnectionSettings.IdleTimeout=30 &Version=2012-06-01 &Action=ModifyLoadBalancerAttributes &AUTHPARAMS my-test-loadbalancer 30 83c88b9d-12b7-11e3-8b82-87b12EXAMPLE.

Documentation

data ModifyLoadBalancerAttributes

Constructors

ModifyLoadBalancerAttributes

Fields

_mlbaiLoadBalancerName :: Text

The name of the load balancer.

_mlbaiLoadBalancerAttributes :: LoadBalancerAttributes

Attributes of the load balancer.

mlbaiLoadBalancerName :: Lens' ModifyLoadBalancerAttributes Text

mlbaiLoadBalancerAttributes :: Lens' ModifyLoadBalancerAttributes LoadBalancerAttributes

data ModifyLoadBalancerAttributesResponse

Constructors

ModifyLoadBalancerAttributesResponse

Fields

_mlbaoLoadBalancerName :: Maybe Text

The name of the load balancer.

_mlbaoLoadBalancerAttributes :: Maybe LoadBalancerAttributes

The LoadBalancerAttributes data type.

mlbaoLoadBalancerName :: Lens' ModifyLoadBalancerAttributesResponse (Maybe Text)

mlbaoLoadBalancerAttributes :: Lens' ModifyLoadBalancerAttributesResponse (Maybe LoadBalancerAttributes)