Network.AWS.ELBV2.AddListenerCertificates

Description

Adds the specified SSL server certificate to the certificate list for the specified HTTPS or TLS listener.

If the certificate in already in the certificate list, the call is successful but the certificate is not added again.

For more information, see HTTPS listeners in the Application Load Balancers Guide or TLS listeners in the Network Load Balancers Guide.

Creating a Request

data AddListenerCertificates

See: newAddListenerCertificates smart constructor.

Constructors

AddListenerCertificates'

Fields

listenerArn :: Text

The Amazon Resource Name (ARN) of the listener.

certificates :: [Certificate]

The certificate to add. You can specify one certificate per call. Set CertificateArn to the certificate ARN but do not set IsDefault.

newAddListenerCertificates

Arguments

:: Text
$sel:listenerArn:AddListenerCertificates'

-> AddListenerCertificates

Create a value of AddListenerCertificates with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:listenerArn:AddListenerCertificates', addListenerCertificates_listenerArn - The Amazon Resource Name (ARN) of the listener.

$sel:certificates:AddListenerCertificates', addListenerCertificates_certificates - The certificate to add. You can specify one certificate per call. Set CertificateArn to the certificate ARN but do not set IsDefault.

Request Lenses

addListenerCertificates_listenerArn :: Lens' AddListenerCertificates Text

The Amazon Resource Name (ARN) of the listener.

addListenerCertificates_certificates :: Lens' AddListenerCertificates [Certificate]

The certificate to add. You can specify one certificate per call. Set CertificateArn to the certificate ARN but do not set IsDefault.

Destructuring the Response

data AddListenerCertificatesResponse

See: newAddListenerCertificatesResponse smart constructor.

Constructors

AddListenerCertificatesResponse'

Fields

certificates :: Maybe [Certificate]

Information about the certificates in the certificate list.

httpStatus :: Int

The response's http status code.

newAddListenerCertificatesResponse

Arguments

:: Int
$sel:httpStatus:AddListenerCertificatesResponse'

-> AddListenerCertificatesResponse

Create a value of AddListenerCertificatesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:certificates:AddListenerCertificates', addListenerCertificatesResponse_certificates - Information about the certificates in the certificate list.

$sel:httpStatus:AddListenerCertificatesResponse', addListenerCertificatesResponse_httpStatus - The response's http status code.

Response Lenses

addListenerCertificatesResponse_certificates :: Lens' AddListenerCertificatesResponse (Maybe [Certificate])

Information about the certificates in the certificate list.

addListenerCertificatesResponse_httpStatus :: Lens' AddListenerCertificatesResponse Int

The response's http status code.