Network.AWS.ELBV2.CreateTargetGroup

Description

Creates a target group.

For more information, see the following:

Target groups for your Application Load Balancers

Target groups for your Network Load Balancers

Target groups for your Gateway Load Balancers

This operation is idempotent, which means that it completes at most one time. If you attempt to create multiple target groups with the same settings, each call succeeds.

Creating a Request

data CreateTargetGroup

See: newCreateTargetGroup smart constructor.

Constructors

CreateTargetGroup'

Fields

protocolVersion :: Maybe Text

HTTP/HTTPS protocol

The protocol version. Specify GRPC to send requests to targets using gRPC. Specify HTTP2 to send requests to targets using HTTP/2. The default is HTTP1, which sends requests to targets using HTTP/1.1.

matcher :: Maybe Matcher

HTTP/HTTPS health checks

The HTTP or gRPC codes to use when checking for a successful response from a target.

healthCheckPath :: Maybe Text

HTTP/HTTPS health checks

The destination for health checks on the targets.

HTTP1 or HTTP2 protocol version

The ping path. The default is /.

GRPC protocol version

The path of a custom health check method with the format /package.service/method. The default is /Amazon Web Services.ALB/healthcheck.

healthCheckEnabled :: Maybe Bool

Indicates whether health checks are enabled. If the target type is lambda, health checks are disabled by default but can be enabled. If the target type is instance, ip, or alb, health checks are always enabled and cannot be disabled.

unhealthyThresholdCount :: Maybe Natural

The number of consecutive health check failures required before considering a target unhealthy. If the target group protocol is HTTP or HTTPS, the default is 2. If the target group protocol is TCP or TLS, this value must be the same as the healthy threshold count. If the target group protocol is GENEVE, the default is 3. If the target type is lambda, the default is 2.

vpcId :: Maybe Text

The identifier of the virtual private cloud (VPC). If the target is a Lambda function, this parameter does not apply. Otherwise, this parameter is required.

protocol :: Maybe ProtocolEnum

The protocol to use for routing traffic to the targets. For Application Load Balancers, the supported protocols are HTTP and HTTPS. For Network Load Balancers, the supported protocols are TCP, TLS, UDP, or TCP_UDP. For Gateway Load Balancers, the supported protocol is GENEVE. A TCP_UDP listener must be associated with a TCP_UDP target group. If the target is a Lambda function, this parameter does not apply.

healthCheckIntervalSeconds :: Maybe Natural

The approximate amount of time, in seconds, between health checks of an individual target. If the target group protocol is TCP, TLS, UDP, or TCP_UDP, the supported values are 10 and 30 seconds. If the target group protocol is HTTP or HTTPS, the default is 30 seconds. If the target group protocol is GENEVE, the default is 10 seconds. If the target type is lambda, the default is 35 seconds.

targetType :: Maybe TargetTypeEnum

The type of target that you must specify when registering targets with this target group. You can't specify targets for a target group using more than one target type.

instance - Register targets by instance ID. This is the default value.

ip - Register targets by IP address. You can specify IP addresses from the subnets of the virtual private cloud (VPC) for the target group, the RFC 1918 range (10.0.0.0/8, 172.16.0.0/12, and 192.168.0.0/16), and the RFC 6598 range (100.64.0.0/10). You can't specify publicly routable IP addresses.

lambda - Register a single Lambda function as a target.

alb - Register a single Application Load Balancer as a target.

healthyThresholdCount :: Maybe Natural

The number of consecutive health checks successes required before considering an unhealthy target healthy. For target groups with a protocol of HTTP or HTTPS, the default is 5. For target groups with a protocol of TCP, TLS, or GENEVE, the default is 3. If the target type is lambda, the default is 5.

healthCheckProtocol :: Maybe ProtocolEnum

The protocol the load balancer uses when performing health checks on targets. For Application Load Balancers, the default is HTTP. For Network Load Balancers and Gateway Load Balancers, the default is TCP. The TCP protocol is not supported for health checks if the protocol of the target group is HTTP or HTTPS. The GENEVE, TLS, UDP, and TCP_UDP protocols are not supported for health checks.

ipAddressType :: Maybe TargetGroupIpAddressTypeEnum

The type of IP address used for this target group. The possible values are ipv4 and ipv6. This is an optional parameter. If not specified, the IP address type defaults to ipv4.

healthCheckTimeoutSeconds :: Maybe Natural

The amount of time, in seconds, during which no response from a target means a failed health check. For target groups with a protocol of HTTP, HTTPS, or GENEVE, the default is 5 seconds. For target groups with a protocol of TCP or TLS, this value must be 6 seconds for HTTP health checks and 10 seconds for TCP and HTTPS health checks. If the target type is lambda, the default is 30 seconds.

healthCheckPort :: Maybe Text

The port the load balancer uses when performing health checks on targets. If the protocol is HTTP, HTTPS, TCP, TLS, UDP, or TCP_UDP, the default is traffic-port, which is the port on which each target receives traffic from the load balancer. If the protocol is GENEVE, the default is port 80.

tags :: Maybe (NonEmpty Tag)

The tags to assign to the target group.

port :: Maybe Natural

The port on which the targets receive traffic. This port is used unless you specify a port override when registering the target. If the target is a Lambda function, this parameter does not apply. If the protocol is GENEVE, the supported port is 6081.

name :: Text

The name of the target group.

This name must be unique per region per account, can have a maximum of 32 characters, must contain only alphanumeric characters or hyphens, and must not begin or end with a hyphen.

newCreateTargetGroup

Arguments

:: Text
$sel:name:CreateTargetGroup'

-> CreateTargetGroup

Create a value of CreateTargetGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:protocolVersion:CreateTargetGroup', createTargetGroup_protocolVersion - [HTTP/HTTPS protocol] The protocol version. Specify GRPC to send requests to targets using gRPC. Specify HTTP2 to send requests to targets using HTTP/2. The default is HTTP1, which sends requests to targets using HTTP/1.1.

$sel:matcher:CreateTargetGroup', createTargetGroup_matcher - [HTTP/HTTPS health checks] The HTTP or gRPC codes to use when checking for a successful response from a target.

$sel:healthCheckPath:CreateTargetGroup', createTargetGroup_healthCheckPath - [HTTP/HTTPS health checks] The destination for health checks on the targets.

HTTP1 or HTTP2 protocol version

The ping path. The default is /.

GRPC protocol version

The path of a custom health check method with the format /package.service/method. The default is /Amazon Web Services.ALB/healthcheck.

$sel:healthCheckEnabled:CreateTargetGroup', createTargetGroup_healthCheckEnabled - Indicates whether health checks are enabled. If the target type is lambda, health checks are disabled by default but can be enabled. If the target type is instance, ip, or alb, health checks are always enabled and cannot be disabled.

$sel:unhealthyThresholdCount:CreateTargetGroup', createTargetGroup_unhealthyThresholdCount - The number of consecutive health check failures required before considering a target unhealthy. If the target group protocol is HTTP or HTTPS, the default is 2. If the target group protocol is TCP or TLS, this value must be the same as the healthy threshold count. If the target group protocol is GENEVE, the default is 3. If the target type is lambda, the default is 2.

$sel:vpcId:CreateTargetGroup', createTargetGroup_vpcId - The identifier of the virtual private cloud (VPC). If the target is a Lambda function, this parameter does not apply. Otherwise, this parameter is required.

$sel:protocol:CreateTargetGroup', createTargetGroup_protocol - The protocol to use for routing traffic to the targets. For Application Load Balancers, the supported protocols are HTTP and HTTPS. For Network Load Balancers, the supported protocols are TCP, TLS, UDP, or TCP_UDP. For Gateway Load Balancers, the supported protocol is GENEVE. A TCP_UDP listener must be associated with a TCP_UDP target group. If the target is a Lambda function, this parameter does not apply.

$sel:healthCheckIntervalSeconds:CreateTargetGroup', createTargetGroup_healthCheckIntervalSeconds - The approximate amount of time, in seconds, between health checks of an individual target. If the target group protocol is TCP, TLS, UDP, or TCP_UDP, the supported values are 10 and 30 seconds. If the target group protocol is HTTP or HTTPS, the default is 30 seconds. If the target group protocol is GENEVE, the default is 10 seconds. If the target type is lambda, the default is 35 seconds.

$sel:targetType:CreateTargetGroup', createTargetGroup_targetType - The type of target that you must specify when registering targets with this target group. You can't specify targets for a target group using more than one target type.

instance - Register targets by instance ID. This is the default value.

ip - Register targets by IP address. You can specify IP addresses from the subnets of the virtual private cloud (VPC) for the target group, the RFC 1918 range (10.0.0.0/8, 172.16.0.0/12, and 192.168.0.0/16), and the RFC 6598 range (100.64.0.0/10). You can't specify publicly routable IP addresses.

lambda - Register a single Lambda function as a target.

alb - Register a single Application Load Balancer as a target.

$sel:healthyThresholdCount:CreateTargetGroup', createTargetGroup_healthyThresholdCount - The number of consecutive health checks successes required before considering an unhealthy target healthy. For target groups with a protocol of HTTP or HTTPS, the default is 5. For target groups with a protocol of TCP, TLS, or GENEVE, the default is 3. If the target type is lambda, the default is 5.

$sel:healthCheckProtocol:CreateTargetGroup', createTargetGroup_healthCheckProtocol - The protocol the load balancer uses when performing health checks on targets. For Application Load Balancers, the default is HTTP. For Network Load Balancers and Gateway Load Balancers, the default is TCP. The TCP protocol is not supported for health checks if the protocol of the target group is HTTP or HTTPS. The GENEVE, TLS, UDP, and TCP_UDP protocols are not supported for health checks.

$sel:ipAddressType:CreateTargetGroup', createTargetGroup_ipAddressType - The type of IP address used for this target group. The possible values are ipv4 and ipv6. This is an optional parameter. If not specified, the IP address type defaults to ipv4.

$sel:healthCheckTimeoutSeconds:CreateTargetGroup', createTargetGroup_healthCheckTimeoutSeconds - The amount of time, in seconds, during which no response from a target means a failed health check. For target groups with a protocol of HTTP, HTTPS, or GENEVE, the default is 5 seconds. For target groups with a protocol of TCP or TLS, this value must be 6 seconds for HTTP health checks and 10 seconds for TCP and HTTPS health checks. If the target type is lambda, the default is 30 seconds.

$sel:healthCheckPort:CreateTargetGroup', createTargetGroup_healthCheckPort - The port the load balancer uses when performing health checks on targets. If the protocol is HTTP, HTTPS, TCP, TLS, UDP, or TCP_UDP, the default is traffic-port, which is the port on which each target receives traffic from the load balancer. If the protocol is GENEVE, the default is port 80.

$sel:tags:CreateTargetGroup', createTargetGroup_tags - The tags to assign to the target group.

$sel:port:CreateTargetGroup', createTargetGroup_port - The port on which the targets receive traffic. This port is used unless you specify a port override when registering the target. If the target is a Lambda function, this parameter does not apply. If the protocol is GENEVE, the supported port is 6081.

$sel:name:CreateTargetGroup', createTargetGroup_name - The name of the target group.

This name must be unique per region per account, can have a maximum of 32 characters, must contain only alphanumeric characters or hyphens, and must not begin or end with a hyphen.

Request Lenses

createTargetGroup_protocolVersion :: Lens' CreateTargetGroup (Maybe Text)

HTTP/HTTPS protocol

The protocol version. Specify GRPC to send requests to targets using gRPC. Specify HTTP2 to send requests to targets using HTTP/2. The default is HTTP1, which sends requests to targets using HTTP/1.1.

createTargetGroup_matcher :: Lens' CreateTargetGroup (Maybe Matcher)

HTTP/HTTPS health checks

The HTTP or gRPC codes to use when checking for a successful response from a target.

createTargetGroup_healthCheckPath :: Lens' CreateTargetGroup (Maybe Text)

HTTP/HTTPS health checks

The destination for health checks on the targets.

HTTP1 or HTTP2 protocol version

The ping path. The default is /.

GRPC protocol version

The path of a custom health check method with the format /package.service/method. The default is /Amazon Web Services.ALB/healthcheck.

createTargetGroup_healthCheckEnabled :: Lens' CreateTargetGroup (Maybe Bool)

Indicates whether health checks are enabled. If the target type is lambda, health checks are disabled by default but can be enabled. If the target type is instance, ip, or alb, health checks are always enabled and cannot be disabled.

createTargetGroup_unhealthyThresholdCount :: Lens' CreateTargetGroup (Maybe Natural)

The number of consecutive health check failures required before considering a target unhealthy. If the target group protocol is HTTP or HTTPS, the default is 2. If the target group protocol is TCP or TLS, this value must be the same as the healthy threshold count. If the target group protocol is GENEVE, the default is 3. If the target type is lambda, the default is 2.

createTargetGroup_vpcId :: Lens' CreateTargetGroup (Maybe Text)

The identifier of the virtual private cloud (VPC). If the target is a Lambda function, this parameter does not apply. Otherwise, this parameter is required.

createTargetGroup_protocol :: Lens' CreateTargetGroup (Maybe ProtocolEnum)

The protocol to use for routing traffic to the targets. For Application Load Balancers, the supported protocols are HTTP and HTTPS. For Network Load Balancers, the supported protocols are TCP, TLS, UDP, or TCP_UDP. For Gateway Load Balancers, the supported protocol is GENEVE. A TCP_UDP listener must be associated with a TCP_UDP target group. If the target is a Lambda function, this parameter does not apply.

createTargetGroup_healthCheckIntervalSeconds :: Lens' CreateTargetGroup (Maybe Natural)

The approximate amount of time, in seconds, between health checks of an individual target. If the target group protocol is TCP, TLS, UDP, or TCP_UDP, the supported values are 10 and 30 seconds. If the target group protocol is HTTP or HTTPS, the default is 30 seconds. If the target group protocol is GENEVE, the default is 10 seconds. If the target type is lambda, the default is 35 seconds.

createTargetGroup_targetType :: Lens' CreateTargetGroup (Maybe TargetTypeEnum)

The type of target that you must specify when registering targets with this target group. You can't specify targets for a target group using more than one target type.

instance - Register targets by instance ID. This is the default value.

ip - Register targets by IP address. You can specify IP addresses from the subnets of the virtual private cloud (VPC) for the target group, the RFC 1918 range (10.0.0.0/8, 172.16.0.0/12, and 192.168.0.0/16), and the RFC 6598 range (100.64.0.0/10). You can't specify publicly routable IP addresses.

lambda - Register a single Lambda function as a target.

alb - Register a single Application Load Balancer as a target.

createTargetGroup_healthyThresholdCount :: Lens' CreateTargetGroup (Maybe Natural)

The number of consecutive health checks successes required before considering an unhealthy target healthy. For target groups with a protocol of HTTP or HTTPS, the default is 5. For target groups with a protocol of TCP, TLS, or GENEVE, the default is 3. If the target type is lambda, the default is 5.

createTargetGroup_healthCheckProtocol :: Lens' CreateTargetGroup (Maybe ProtocolEnum)

The protocol the load balancer uses when performing health checks on targets. For Application Load Balancers, the default is HTTP. For Network Load Balancers and Gateway Load Balancers, the default is TCP. The TCP protocol is not supported for health checks if the protocol of the target group is HTTP or HTTPS. The GENEVE, TLS, UDP, and TCP_UDP protocols are not supported for health checks.

createTargetGroup_ipAddressType :: Lens' CreateTargetGroup (Maybe TargetGroupIpAddressTypeEnum)

The type of IP address used for this target group. The possible values are ipv4 and ipv6. This is an optional parameter. If not specified, the IP address type defaults to ipv4.

createTargetGroup_healthCheckTimeoutSeconds :: Lens' CreateTargetGroup (Maybe Natural)

The amount of time, in seconds, during which no response from a target means a failed health check. For target groups with a protocol of HTTP, HTTPS, or GENEVE, the default is 5 seconds. For target groups with a protocol of TCP or TLS, this value must be 6 seconds for HTTP health checks and 10 seconds for TCP and HTTPS health checks. If the target type is lambda, the default is 30 seconds.

createTargetGroup_healthCheckPort :: Lens' CreateTargetGroup (Maybe Text)

The port the load balancer uses when performing health checks on targets. If the protocol is HTTP, HTTPS, TCP, TLS, UDP, or TCP_UDP, the default is traffic-port, which is the port on which each target receives traffic from the load balancer. If the protocol is GENEVE, the default is port 80.

createTargetGroup_tags :: Lens' CreateTargetGroup (Maybe (NonEmpty Tag))

The tags to assign to the target group.

createTargetGroup_port :: Lens' CreateTargetGroup (Maybe Natural)

The port on which the targets receive traffic. This port is used unless you specify a port override when registering the target. If the target is a Lambda function, this parameter does not apply. If the protocol is GENEVE, the supported port is 6081.

createTargetGroup_name :: Lens' CreateTargetGroup Text

The name of the target group.

This name must be unique per region per account, can have a maximum of 32 characters, must contain only alphanumeric characters or hyphens, and must not begin or end with a hyphen.

Destructuring the Response

data CreateTargetGroupResponse

See: newCreateTargetGroupResponse smart constructor.

Constructors

CreateTargetGroupResponse'

Fields

targetGroups :: Maybe [TargetGroup]

Information about the target group.

httpStatus :: Int

The response's http status code.

newCreateTargetGroupResponse

Arguments

:: Int
$sel:httpStatus:CreateTargetGroupResponse'

-> CreateTargetGroupResponse

Create a value of CreateTargetGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetGroups:CreateTargetGroupResponse', createTargetGroupResponse_targetGroups - Information about the target group.

$sel:httpStatus:CreateTargetGroupResponse', createTargetGroupResponse_httpStatus - The response's http status code.

Response Lenses

createTargetGroupResponse_targetGroups :: Lens' CreateTargetGroupResponse (Maybe [TargetGroup])

Information about the target group.

createTargetGroupResponse_httpStatus :: Lens' CreateTargetGroupResponse Int

The response's http status code.