Network.AWS.ELBV2.Types.RuleCondition

Description

Documentation

data RuleCondition

Information about a condition for a rule.

Each rule can optionally include up to one of each of the following conditions: http-request-method, host-header, path-pattern, and source-ip. Each rule can also optionally include one or more of each of the following conditions: http-header and query-string.

See: newRuleCondition smart constructor.

Constructors

RuleCondition'

Fields

field :: Maybe Text

The field in the HTTP request. The following are the possible values:

http-header

http-request-method

host-header

path-pattern

query-string

source-ip

httpHeaderConfig :: Maybe HttpHeaderConditionConfig

Information for an HTTP header condition. Specify only when Field is http-header.

hostHeaderConfig :: Maybe HostHeaderConditionConfig

Information for a host header condition. Specify only when Field is host-header.

values :: Maybe [Text]

The condition value. Specify only when Field is host-header or path-pattern. Alternatively, to specify multiple host names or multiple path patterns, use HostHeaderConfig or PathPatternConfig.

If Field is host-header and you are not using HostHeaderConfig, you can specify a single host name (for example, my.example.com) in Values. A host name is case insensitive, can be up to 128 characters in length, and can contain any of the following characters.

A-Z, a-z, 0-9

- .

* (matches 0 or more characters)

? (matches exactly 1 character)

If Field is path-pattern and you are not using PathPatternConfig, you can specify a single path pattern (for example, /img/*) in Values. A path pattern is case-sensitive, can be up to 128 characters in length, and can contain any of the following characters.

A-Z, a-z, 0-9

_ - . $ / ~ " ' @ : +

& (using &amp;)

* (matches 0 or more characters)

? (matches exactly 1 character)

sourceIpConfig :: Maybe SourceIpConditionConfig

Information for a source IP condition. Specify only when Field is source-ip.

httpRequestMethodConfig :: Maybe HttpRequestMethodConditionConfig

Information for an HTTP method condition. Specify only when Field is http-request-method.

pathPatternConfig :: Maybe PathPatternConditionConfig

Information for a path pattern condition. Specify only when Field is path-pattern.

queryStringConfig :: Maybe QueryStringConditionConfig

Information for a query string condition. Specify only when Field is query-string.

newRuleCondition :: RuleCondition

Create a value of RuleCondition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:field:RuleCondition', ruleCondition_field - The field in the HTTP request. The following are the possible values:

http-header

http-request-method

host-header

path-pattern

query-string

source-ip

$sel:httpHeaderConfig:RuleCondition', ruleCondition_httpHeaderConfig - Information for an HTTP header condition. Specify only when Field is http-header.

$sel:hostHeaderConfig:RuleCondition', ruleCondition_hostHeaderConfig - Information for a host header condition. Specify only when Field is host-header.

$sel:values:RuleCondition', ruleCondition_values - The condition value. Specify only when Field is host-header or path-pattern. Alternatively, to specify multiple host names or multiple path patterns, use HostHeaderConfig or PathPatternConfig.

If Field is host-header and you are not using HostHeaderConfig, you can specify a single host name (for example, my.example.com) in Values. A host name is case insensitive, can be up to 128 characters in length, and can contain any of the following characters.

A-Z, a-z, 0-9

- .

* (matches 0 or more characters)

? (matches exactly 1 character)

If Field is path-pattern and you are not using PathPatternConfig, you can specify a single path pattern (for example, /img/*) in Values. A path pattern is case-sensitive, can be up to 128 characters in length, and can contain any of the following characters.

A-Z, a-z, 0-9

_ - . $ / ~ " ' @ : +

& (using &amp;)

* (matches 0 or more characters)

? (matches exactly 1 character)

$sel:sourceIpConfig:RuleCondition', ruleCondition_sourceIpConfig - Information for a source IP condition. Specify only when Field is source-ip.

$sel:httpRequestMethodConfig:RuleCondition', ruleCondition_httpRequestMethodConfig - Information for an HTTP method condition. Specify only when Field is http-request-method.

$sel:pathPatternConfig:RuleCondition', ruleCondition_pathPatternConfig - Information for a path pattern condition. Specify only when Field is path-pattern.

$sel:queryStringConfig:RuleCondition', ruleCondition_queryStringConfig - Information for a query string condition. Specify only when Field is query-string.

ruleCondition_field :: Lens' RuleCondition (Maybe Text)

The field in the HTTP request. The following are the possible values:

http-header

http-request-method

host-header

path-pattern

query-string

source-ip

ruleCondition_httpHeaderConfig :: Lens' RuleCondition (Maybe HttpHeaderConditionConfig)

Information for an HTTP header condition. Specify only when Field is http-header.

ruleCondition_hostHeaderConfig :: Lens' RuleCondition (Maybe HostHeaderConditionConfig)

Information for a host header condition. Specify only when Field is host-header.

ruleCondition_values :: Lens' RuleCondition (Maybe [Text])

The condition value. Specify only when Field is host-header or path-pattern. Alternatively, to specify multiple host names or multiple path patterns, use HostHeaderConfig or PathPatternConfig.

If Field is host-header and you are not using HostHeaderConfig, you can specify a single host name (for example, my.example.com) in Values. A host name is case insensitive, can be up to 128 characters in length, and can contain any of the following characters.

A-Z, a-z, 0-9

- .

* (matches 0 or more characters)

? (matches exactly 1 character)

If Field is path-pattern and you are not using PathPatternConfig, you can specify a single path pattern (for example, /img/*) in Values. A path pattern is case-sensitive, can be up to 128 characters in length, and can contain any of the following characters.

A-Z, a-z, 0-9

_ - . $ / ~ " ' @ : +

& (using &amp;)

* (matches 0 or more characters)

? (matches exactly 1 character)

ruleCondition_sourceIpConfig :: Lens' RuleCondition (Maybe SourceIpConditionConfig)

Information for a source IP condition. Specify only when Field is source-ip.

ruleCondition_httpRequestMethodConfig :: Lens' RuleCondition (Maybe HttpRequestMethodConditionConfig)

Information for an HTTP method condition. Specify only when Field is http-request-method.

ruleCondition_pathPatternConfig :: Lens' RuleCondition (Maybe PathPatternConditionConfig)

Information for a path pattern condition. Specify only when Field is path-pattern.

ruleCondition_queryStringConfig :: Lens' RuleCondition (Maybe QueryStringConditionConfig)

Information for a query string condition. Specify only when Field is query-string.