Network.AWS.ELBv2.CreateListener

Description

Creates a listener for the specified Application Load Balancer, Network Load Balancer, or Gateway Load Balancer.

For more information, see the following:

Listeners for your Application Load Balancers

Listeners for your Network Load Balancers

Listeners for your Gateway Load Balancers

This operation is idempotent, which means that it completes at most one time. If you attempt to create multiple listeners with the same settings, each call succeeds.

Creating a Request

data CreateListener

See: newCreateListener smart constructor.

Constructors

CreateListener'

Fields

sslPolicy :: Maybe Text

HTTPS and TLS listeners

The security policy that defines which protocols and ciphers are supported.

For more information, see Security policies in the Application Load Balancers Guide and Security policies in the Network Load Balancers Guide.

tags :: Maybe (NonEmpty Tag)

The tags to assign to the listener.

port :: Maybe Natural

The port on which the load balancer is listening. You cannot specify a port for a Gateway Load Balancer.

certificates :: Maybe [Certificate]

HTTPS and TLS listeners

The default certificate for the listener. You must provide exactly one certificate. Set CertificateArn to the certificate ARN but do not set IsDefault.

protocol :: Maybe ProtocolEnum

The protocol for connections from clients to the load balancer. For Application Load Balancers, the supported protocols are HTTP and HTTPS. For Network Load Balancers, the supported protocols are TCP, TLS, UDP, and TCP_UDP. You can’t specify the UDP or TCP_UDP protocol if dual-stack mode is enabled. You cannot specify a protocol for a Gateway Load Balancer.

alpnPolicy :: Maybe [Text]

TLS listeners

The name of the Application-Layer Protocol Negotiation (ALPN) policy. You can specify one policy name. The following are the possible values:

HTTP1Only

HTTP2Only

HTTP2Optional

HTTP2Preferred

None

For more information, see ALPN policies in the Network Load Balancers Guide.

loadBalancerArn :: Text

The Amazon Resource Name (ARN) of the load balancer.

defaultActions :: [Action]

The actions for the default rule.

newCreateListener

Arguments

:: Text
$sel:loadBalancerArn:CreateListener'

-> CreateListener

Create a value of CreateListener with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sslPolicy:CreateListener', createListener_sslPolicy - [HTTPS and TLS listeners] The security policy that defines which protocols and ciphers are supported.

For more information, see Security policies in the Application Load Balancers Guide and Security policies in the Network Load Balancers Guide.

$sel:tags:CreateListener', createListener_tags - The tags to assign to the listener.

$sel:port:CreateListener', createListener_port - The port on which the load balancer is listening. You cannot specify a port for a Gateway Load Balancer.

$sel:certificates:CreateListener', createListener_certificates - [HTTPS and TLS listeners] The default certificate for the listener. You must provide exactly one certificate. Set CertificateArn to the certificate ARN but do not set IsDefault.

$sel:protocol:CreateListener', createListener_protocol - The protocol for connections from clients to the load balancer. For Application Load Balancers, the supported protocols are HTTP and HTTPS. For Network Load Balancers, the supported protocols are TCP, TLS, UDP, and TCP_UDP. You can’t specify the UDP or TCP_UDP protocol if dual-stack mode is enabled. You cannot specify a protocol for a Gateway Load Balancer.

$sel:alpnPolicy:CreateListener', createListener_alpnPolicy - [TLS listeners] The name of the Application-Layer Protocol Negotiation (ALPN) policy. You can specify one policy name. The following are the possible values:

HTTP1Only

HTTP2Only

HTTP2Optional

HTTP2Preferred

None

For more information, see ALPN policies in the Network Load Balancers Guide.

$sel:loadBalancerArn:CreateListener', createListener_loadBalancerArn - The Amazon Resource Name (ARN) of the load balancer.

$sel:defaultActions:CreateListener', createListener_defaultActions - The actions for the default rule.

Request Lenses

createListener_sslPolicy :: Lens' CreateListener (Maybe Text)

HTTPS and TLS listeners

The security policy that defines which protocols and ciphers are supported.

For more information, see Security policies in the Application Load Balancers Guide and Security policies in the Network Load Balancers Guide.

createListener_tags :: Lens' CreateListener (Maybe (NonEmpty Tag))

The tags to assign to the listener.

createListener_port :: Lens' CreateListener (Maybe Natural)

The port on which the load balancer is listening. You cannot specify a port for a Gateway Load Balancer.

createListener_certificates :: Lens' CreateListener (Maybe [Certificate])

HTTPS and TLS listeners

The default certificate for the listener. You must provide exactly one certificate. Set CertificateArn to the certificate ARN but do not set IsDefault.

createListener_protocol :: Lens' CreateListener (Maybe ProtocolEnum)

The protocol for connections from clients to the load balancer. For Application Load Balancers, the supported protocols are HTTP and HTTPS. For Network Load Balancers, the supported protocols are TCP, TLS, UDP, and TCP_UDP. You can’t specify the UDP or TCP_UDP protocol if dual-stack mode is enabled. You cannot specify a protocol for a Gateway Load Balancer.

createListener_alpnPolicy :: Lens' CreateListener (Maybe [Text])

TLS listeners

The name of the Application-Layer Protocol Negotiation (ALPN) policy. You can specify one policy name. The following are the possible values:

HTTP1Only

HTTP2Only

HTTP2Optional

HTTP2Preferred

None

For more information, see ALPN policies in the Network Load Balancers Guide.

createListener_loadBalancerArn :: Lens' CreateListener Text

The Amazon Resource Name (ARN) of the load balancer.

createListener_defaultActions :: Lens' CreateListener [Action]

The actions for the default rule.

Destructuring the Response

data CreateListenerResponse

See: newCreateListenerResponse smart constructor.

Constructors

CreateListenerResponse'

Fields

listeners :: Maybe [Listener]

Information about the listener.

httpStatus :: Int

The response's http status code.

newCreateListenerResponse

Arguments

:: Int
$sel:httpStatus:CreateListenerResponse'

-> CreateListenerResponse

Create a value of CreateListenerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:listeners:CreateListenerResponse', createListenerResponse_listeners - Information about the listener.

$sel:httpStatus:CreateListenerResponse', createListenerResponse_httpStatus - The response's http status code.

Response Lenses

createListenerResponse_listeners :: Lens' CreateListenerResponse (Maybe [Listener])

Information about the listener.

createListenerResponse_httpStatus :: Lens' CreateListenerResponse Int

The response's http status code.