Network.AWS.ELBv2.CreateRule

Description

Creates a rule for the specified listener. The listener must be associated with an Application Load Balancer.

Each rule consists of a priority, one or more actions, and one or more conditions. Rules are evaluated in priority order, from the lowest value to the highest value. When the conditions for a rule are met, its actions are performed. If the conditions for no rules are met, the actions for the default rule are performed. For more information, see Listener rules in the Application Load Balancers Guide.

Creating a Request

data CreateRule

See: newCreateRule smart constructor.

Constructors

CreateRule'

Fields

tags :: Maybe (NonEmpty Tag)

The tags to assign to the rule.

listenerArn :: Text

The Amazon Resource Name (ARN) of the listener.

conditions :: [RuleCondition]

The conditions.

priority :: Natural

The rule priority. A listener can't have multiple rules with the same priority.

actions :: [Action]

The actions.

newCreateRule

Arguments

:: Text
$sel:listenerArn:CreateRule'

-> Natural
$sel:priority:CreateRule'

-> CreateRule

Create a value of CreateRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateRule', createRule_tags - The tags to assign to the rule.

$sel:listenerArn:CreateRule', createRule_listenerArn - The Amazon Resource Name (ARN) of the listener.

$sel:conditions:CreateRule', createRule_conditions - The conditions.

$sel:priority:CreateRule', createRule_priority - The rule priority. A listener can't have multiple rules with the same priority.

$sel:actions:CreateRule', createRule_actions - The actions.

Request Lenses

createRule_tags :: Lens' CreateRule (Maybe (NonEmpty Tag))

The tags to assign to the rule.

createRule_listenerArn :: Lens' CreateRule Text

The Amazon Resource Name (ARN) of the listener.

createRule_conditions :: Lens' CreateRule [RuleCondition]

The conditions.

createRule_priority :: Lens' CreateRule Natural

The rule priority. A listener can't have multiple rules with the same priority.

createRule_actions :: Lens' CreateRule [Action]

The actions.

Destructuring the Response

data CreateRuleResponse

See: newCreateRuleResponse smart constructor.

Constructors

CreateRuleResponse'

Fields

rules :: Maybe [Rule]

Information about the rule.

httpStatus :: Int

The response's http status code.

newCreateRuleResponse

Arguments

:: Int
$sel:httpStatus:CreateRuleResponse'

-> CreateRuleResponse

Create a value of CreateRuleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:rules:CreateRuleResponse', createRuleResponse_rules - Information about the rule.

$sel:httpStatus:CreateRuleResponse', createRuleResponse_httpStatus - The response's http status code.

Response Lenses

createRuleResponse_rules :: Lens' CreateRuleResponse (Maybe [Rule])

Information about the rule.

createRuleResponse_httpStatus :: Lens' CreateRuleResponse Int

The response's http status code.