Network.AWS.ELBv2.DescribeListeners

Description

Describes the specified listeners or the listeners for the specified Application Load Balancer, Network Load Balancer, or Gateway Load Balancer. You must specify either a load balancer or one or more listeners.

This operation returns paginated results.

Creating a Request

data DescribeListeners

See: newDescribeListeners smart constructor.

Constructors

DescribeListeners'

Fields

loadBalancerArn :: Maybe Text

The Amazon Resource Name (ARN) of the load balancer.

pageSize :: Maybe Natural

The maximum number of results to return with this call.

listenerArns :: Maybe [Text]

The Amazon Resource Names (ARN) of the listeners.

marker :: Maybe Text

The marker for the next set of results. (You received this marker from a previous call.)

newDescribeListeners :: DescribeListeners

Create a value of DescribeListeners with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:loadBalancerArn:DescribeListeners', describeListeners_loadBalancerArn - The Amazon Resource Name (ARN) of the load balancer.

$sel:pageSize:DescribeListeners', describeListeners_pageSize - The maximum number of results to return with this call.

$sel:listenerArns:DescribeListeners', describeListeners_listenerArns - The Amazon Resource Names (ARN) of the listeners.

$sel:marker:DescribeListeners', describeListeners_marker - The marker for the next set of results. (You received this marker from a previous call.)

Request Lenses

describeListeners_loadBalancerArn :: Lens' DescribeListeners (Maybe Text)

The Amazon Resource Name (ARN) of the load balancer.

describeListeners_pageSize :: Lens' DescribeListeners (Maybe Natural)

The maximum number of results to return with this call.

describeListeners_listenerArns :: Lens' DescribeListeners (Maybe [Text])

The Amazon Resource Names (ARN) of the listeners.

describeListeners_marker :: Lens' DescribeListeners (Maybe Text)

The marker for the next set of results. (You received this marker from a previous call.)

Destructuring the Response

data DescribeListenersResponse

See: newDescribeListenersResponse smart constructor.

Constructors

DescribeListenersResponse'

Fields

nextMarker :: Maybe Text

If there are additional results, this is the marker for the next set of results. Otherwise, this is null.

listeners :: Maybe [Listener]

Information about the listeners.

httpStatus :: Int

The response's http status code.

newDescribeListenersResponse

Arguments

:: Int
$sel:httpStatus:DescribeListenersResponse'

-> DescribeListenersResponse

Create a value of DescribeListenersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextMarker:DescribeListenersResponse', describeListenersResponse_nextMarker - If there are additional results, this is the marker for the next set of results. Otherwise, this is null.

$sel:listeners:DescribeListenersResponse', describeListenersResponse_listeners - Information about the listeners.

$sel:httpStatus:DescribeListenersResponse', describeListenersResponse_httpStatus - The response's http status code.

Response Lenses

describeListenersResponse_nextMarker :: Lens' DescribeListenersResponse (Maybe Text)

If there are additional results, this is the marker for the next set of results. Otherwise, this is null.

describeListenersResponse_listeners :: Lens' DescribeListenersResponse (Maybe [Listener])

Information about the listeners.

describeListenersResponse_httpStatus :: Lens' DescribeListenersResponse Int

The response's http status code.