Network.AWS.ELBv2.DescribeTargetGroups

Description

Describes the specified target groups or all of your target groups. By default, all target groups are described. Alternatively, you can specify one of the following to filter the results: the ARN of the load balancer, the names of one or more target groups, or the ARNs of one or more target groups.

This operation returns paginated results.

Creating a Request

data DescribeTargetGroups

See: newDescribeTargetGroups smart constructor.

Constructors

DescribeTargetGroups'

Fields

loadBalancerArn :: Maybe Text

The Amazon Resource Name (ARN) of the load balancer.

names :: Maybe [Text]

The names of the target groups.

pageSize :: Maybe Natural

The maximum number of results to return with this call.

targetGroupArns :: Maybe [Text]

The Amazon Resource Names (ARN) of the target groups.

marker :: Maybe Text

The marker for the next set of results. (You received this marker from a previous call.)

newDescribeTargetGroups :: DescribeTargetGroups

Create a value of DescribeTargetGroups with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:loadBalancerArn:DescribeTargetGroups', describeTargetGroups_loadBalancerArn - The Amazon Resource Name (ARN) of the load balancer.

$sel:names:DescribeTargetGroups', describeTargetGroups_names - The names of the target groups.

$sel:pageSize:DescribeTargetGroups', describeTargetGroups_pageSize - The maximum number of results to return with this call.

$sel:targetGroupArns:DescribeTargetGroups', describeTargetGroups_targetGroupArns - The Amazon Resource Names (ARN) of the target groups.

$sel:marker:DescribeTargetGroups', describeTargetGroups_marker - The marker for the next set of results. (You received this marker from a previous call.)

Request Lenses

describeTargetGroups_loadBalancerArn :: Lens' DescribeTargetGroups (Maybe Text)

The Amazon Resource Name (ARN) of the load balancer.

describeTargetGroups_names :: Lens' DescribeTargetGroups (Maybe [Text])

The names of the target groups.

describeTargetGroups_pageSize :: Lens' DescribeTargetGroups (Maybe Natural)

The maximum number of results to return with this call.

describeTargetGroups_targetGroupArns :: Lens' DescribeTargetGroups (Maybe [Text])

The Amazon Resource Names (ARN) of the target groups.

describeTargetGroups_marker :: Lens' DescribeTargetGroups (Maybe Text)

The marker for the next set of results. (You received this marker from a previous call.)

Destructuring the Response

data DescribeTargetGroupsResponse

See: newDescribeTargetGroupsResponse smart constructor.

Constructors

DescribeTargetGroupsResponse'

Fields

nextMarker :: Maybe Text

If there are additional results, this is the marker for the next set of results. Otherwise, this is null.

targetGroups :: Maybe [TargetGroup]

Information about the target groups.

httpStatus :: Int

The response's http status code.

newDescribeTargetGroupsResponse

Arguments

:: Int
$sel:httpStatus:DescribeTargetGroupsResponse'

-> DescribeTargetGroupsResponse

Create a value of DescribeTargetGroupsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextMarker:DescribeTargetGroupsResponse', describeTargetGroupsResponse_nextMarker - If there are additional results, this is the marker for the next set of results. Otherwise, this is null.

$sel:targetGroups:DescribeTargetGroupsResponse', describeTargetGroupsResponse_targetGroups - Information about the target groups.

$sel:httpStatus:DescribeTargetGroupsResponse', describeTargetGroupsResponse_httpStatus - The response's http status code.

Response Lenses

describeTargetGroupsResponse_nextMarker :: Lens' DescribeTargetGroupsResponse (Maybe Text)

If there are additional results, this is the marker for the next set of results. Otherwise, this is null.

describeTargetGroupsResponse_targetGroups :: Lens' DescribeTargetGroupsResponse (Maybe [TargetGroup])

Information about the target groups.

describeTargetGroupsResponse_httpStatus :: Lens' DescribeTargetGroupsResponse Int

The response's http status code.