Network.AWS.ELBv2.ModifyListener

Description

Replaces the specified properties of the specified listener. Any properties that you do not specify remain unchanged.

Changing the protocol from HTTPS to HTTP, or from TLS to TCP, removes the security policy and default certificate properties. If you change the protocol from HTTP to HTTPS, or from TCP to TLS, you must add the security policy and default certificate properties.

To add an item to a list, remove an item from a list, or update an item in a list, you must provide the entire list. For example, to add an action, specify a list with the current actions plus the new action.

Creating a Request

data ModifyListener

See: newModifyListener smart constructor.

Constructors

ModifyListener'

Fields

sslPolicy :: Maybe Text

HTTPS and TLS listeners

The security policy that defines which protocols and ciphers are supported.

For more information, see Security policies in the Application Load Balancers Guide or Security policies in the Network Load Balancers Guide.

port :: Maybe Natural

The port for connections from clients to the load balancer. You cannot specify a port for a Gateway Load Balancer.

certificates :: Maybe [Certificate]

HTTPS and TLS listeners

The default certificate for the listener. You must provide exactly one certificate. Set CertificateArn to the certificate ARN but do not set IsDefault.

protocol :: Maybe ProtocolEnum

The protocol for connections from clients to the load balancer. Application Load Balancers support the HTTP and HTTPS protocols. Network Load Balancers support the TCP, TLS, UDP, and TCP_UDP protocols. You can’t change the protocol to UDP or TCP_UDP if dual-stack mode is enabled. You cannot specify a protocol for a Gateway Load Balancer.

defaultActions :: Maybe [Action]

The actions for the default rule.

alpnPolicy :: Maybe [Text]

TLS listeners

The name of the Application-Layer Protocol Negotiation (ALPN) policy. You can specify one policy name. The following are the possible values:

HTTP1Only

HTTP2Only

HTTP2Optional

HTTP2Preferred

None

For more information, see ALPN policies in the Network Load Balancers Guide.

listenerArn :: Text

The Amazon Resource Name (ARN) of the listener.

newModifyListener

Arguments

:: Text
$sel:listenerArn:ModifyListener'

-> ModifyListener

Create a value of ModifyListener with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sslPolicy:ModifyListener', modifyListener_sslPolicy - [HTTPS and TLS listeners] The security policy that defines which protocols and ciphers are supported.

For more information, see Security policies in the Application Load Balancers Guide or Security policies in the Network Load Balancers Guide.

$sel:port:ModifyListener', modifyListener_port - The port for connections from clients to the load balancer. You cannot specify a port for a Gateway Load Balancer.

$sel:certificates:ModifyListener', modifyListener_certificates - [HTTPS and TLS listeners] The default certificate for the listener. You must provide exactly one certificate. Set CertificateArn to the certificate ARN but do not set IsDefault.

$sel:protocol:ModifyListener', modifyListener_protocol - The protocol for connections from clients to the load balancer. Application Load Balancers support the HTTP and HTTPS protocols. Network Load Balancers support the TCP, TLS, UDP, and TCP_UDP protocols. You can’t change the protocol to UDP or TCP_UDP if dual-stack mode is enabled. You cannot specify a protocol for a Gateway Load Balancer.

$sel:defaultActions:ModifyListener', modifyListener_defaultActions - The actions for the default rule.

$sel:alpnPolicy:ModifyListener', modifyListener_alpnPolicy - [TLS listeners] The name of the Application-Layer Protocol Negotiation (ALPN) policy. You can specify one policy name. The following are the possible values:

HTTP1Only

HTTP2Only

HTTP2Optional

HTTP2Preferred

None

For more information, see ALPN policies in the Network Load Balancers Guide.

$sel:listenerArn:ModifyListener', modifyListener_listenerArn - The Amazon Resource Name (ARN) of the listener.

Request Lenses

modifyListener_sslPolicy :: Lens' ModifyListener (Maybe Text)

HTTPS and TLS listeners

The security policy that defines which protocols and ciphers are supported.

For more information, see Security policies in the Application Load Balancers Guide or Security policies in the Network Load Balancers Guide.

modifyListener_port :: Lens' ModifyListener (Maybe Natural)

The port for connections from clients to the load balancer. You cannot specify a port for a Gateway Load Balancer.

modifyListener_certificates :: Lens' ModifyListener (Maybe [Certificate])

HTTPS and TLS listeners

The default certificate for the listener. You must provide exactly one certificate. Set CertificateArn to the certificate ARN but do not set IsDefault.

modifyListener_protocol :: Lens' ModifyListener (Maybe ProtocolEnum)

The protocol for connections from clients to the load balancer. Application Load Balancers support the HTTP and HTTPS protocols. Network Load Balancers support the TCP, TLS, UDP, and TCP_UDP protocols. You can’t change the protocol to UDP or TCP_UDP if dual-stack mode is enabled. You cannot specify a protocol for a Gateway Load Balancer.

modifyListener_defaultActions :: Lens' ModifyListener (Maybe [Action])

The actions for the default rule.

modifyListener_alpnPolicy :: Lens' ModifyListener (Maybe [Text])

TLS listeners

The name of the Application-Layer Protocol Negotiation (ALPN) policy. You can specify one policy name. The following are the possible values:

HTTP1Only

HTTP2Only

HTTP2Optional

HTTP2Preferred

None

For more information, see ALPN policies in the Network Load Balancers Guide.

modifyListener_listenerArn :: Lens' ModifyListener Text

The Amazon Resource Name (ARN) of the listener.

Destructuring the Response

data ModifyListenerResponse

See: newModifyListenerResponse smart constructor.

Constructors

ModifyListenerResponse'

Fields

listeners :: Maybe [Listener]

Information about the modified listener.

httpStatus :: Int

The response's http status code.

newModifyListenerResponse

Arguments

:: Int
$sel:httpStatus:ModifyListenerResponse'

-> ModifyListenerResponse

Create a value of ModifyListenerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:listeners:ModifyListenerResponse', modifyListenerResponse_listeners - Information about the modified listener.

$sel:httpStatus:ModifyListenerResponse', modifyListenerResponse_httpStatus - The response's http status code.

Response Lenses

modifyListenerResponse_listeners :: Lens' ModifyListenerResponse (Maybe [Listener])

Information about the modified listener.

modifyListenerResponse_httpStatus :: Lens' ModifyListenerResponse Int

The response's http status code.