Network.AWS.ELBv2.ModifyLoadBalancerAttributes

Description

Modifies the specified attributes of the specified Application Load Balancer, Network Load Balancer, or Gateway Load Balancer.

If any of the specified attributes can't be modified as requested, the call fails. Any existing attributes that you do not modify retain their current values.

Creating a Request

data ModifyLoadBalancerAttributes

See: newModifyLoadBalancerAttributes smart constructor.

Constructors

ModifyLoadBalancerAttributes'

Fields

loadBalancerArn :: Text

The Amazon Resource Name (ARN) of the load balancer.

attributes :: [LoadBalancerAttribute]

The load balancer attributes.

newModifyLoadBalancerAttributes

Arguments

:: Text
$sel:loadBalancerArn:ModifyLoadBalancerAttributes'

-> ModifyLoadBalancerAttributes

Create a value of ModifyLoadBalancerAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:loadBalancerArn:ModifyLoadBalancerAttributes', modifyLoadBalancerAttributes_loadBalancerArn - The Amazon Resource Name (ARN) of the load balancer.

$sel:attributes:ModifyLoadBalancerAttributes', modifyLoadBalancerAttributes_attributes - The load balancer attributes.

Request Lenses

modifyLoadBalancerAttributes_loadBalancerArn :: Lens' ModifyLoadBalancerAttributes Text

The Amazon Resource Name (ARN) of the load balancer.

modifyLoadBalancerAttributes_attributes :: Lens' ModifyLoadBalancerAttributes [LoadBalancerAttribute]

The load balancer attributes.

Destructuring the Response

data ModifyLoadBalancerAttributesResponse

See: newModifyLoadBalancerAttributesResponse smart constructor.

Constructors

ModifyLoadBalancerAttributesResponse'

Fields

attributes :: Maybe [LoadBalancerAttribute]

Information about the load balancer attributes.

httpStatus :: Int

The response's http status code.

newModifyLoadBalancerAttributesResponse

Arguments

:: Int
$sel:httpStatus:ModifyLoadBalancerAttributesResponse'

-> ModifyLoadBalancerAttributesResponse

Create a value of ModifyLoadBalancerAttributesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributes:ModifyLoadBalancerAttributes', modifyLoadBalancerAttributesResponse_attributes - Information about the load balancer attributes.

$sel:httpStatus:ModifyLoadBalancerAttributesResponse', modifyLoadBalancerAttributesResponse_httpStatus - The response's http status code.

Response Lenses

modifyLoadBalancerAttributesResponse_attributes :: Lens' ModifyLoadBalancerAttributesResponse (Maybe [LoadBalancerAttribute])

Information about the load balancer attributes.

modifyLoadBalancerAttributesResponse_httpStatus :: Lens' ModifyLoadBalancerAttributesResponse Int

The response's http status code.