Network.AWS.ELBv2.ModifyRule

Description

Replaces the specified properties of the specified rule. Any properties that you do not specify are unchanged.

To add an item to a list, remove an item from a list, or update an item in a list, you must provide the entire list. For example, to add an action, specify a list with the current actions plus the new action.

Creating a Request

data ModifyRule

See: newModifyRule smart constructor.

Constructors

ModifyRule'

Fields

actions :: Maybe [Action]

The actions.

conditions :: Maybe [RuleCondition]

The conditions.

ruleArn :: Text

The Amazon Resource Name (ARN) of the rule.

newModifyRule

Arguments

:: Text
$sel:ruleArn:ModifyRule'

-> ModifyRule

Create a value of ModifyRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:actions:ModifyRule', modifyRule_actions - The actions.

$sel:conditions:ModifyRule', modifyRule_conditions - The conditions.

$sel:ruleArn:ModifyRule', modifyRule_ruleArn - The Amazon Resource Name (ARN) of the rule.

Request Lenses

modifyRule_actions :: Lens' ModifyRule (Maybe [Action])

The actions.

modifyRule_conditions :: Lens' ModifyRule (Maybe [RuleCondition])

The conditions.

modifyRule_ruleArn :: Lens' ModifyRule Text

The Amazon Resource Name (ARN) of the rule.

Destructuring the Response

data ModifyRuleResponse

See: newModifyRuleResponse smart constructor.

Constructors

ModifyRuleResponse'

Fields

rules :: Maybe [Rule]

Information about the modified rule.

httpStatus :: Int

The response's http status code.

newModifyRuleResponse

Arguments

:: Int
$sel:httpStatus:ModifyRuleResponse'

-> ModifyRuleResponse

Create a value of ModifyRuleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:rules:ModifyRuleResponse', modifyRuleResponse_rules - Information about the modified rule.

$sel:httpStatus:ModifyRuleResponse', modifyRuleResponse_httpStatus - The response's http status code.

Response Lenses

modifyRuleResponse_rules :: Lens' ModifyRuleResponse (Maybe [Rule])

Information about the modified rule.

modifyRuleResponse_httpStatus :: Lens' ModifyRuleResponse Int

The response's http status code.