Network.AWS.ELBv2.ModifyTargetGroup

Description

Modifies the health checks used when evaluating the health state of the targets in the specified target group.

Creating a Request

data ModifyTargetGroup

See: newModifyTargetGroup smart constructor.

Constructors

ModifyTargetGroup'

Fields

healthCheckEnabled :: Maybe Bool

Indicates whether health checks are enabled.

healthCheckProtocol :: Maybe ProtocolEnum

The protocol the load balancer uses when performing health checks on targets. For Application Load Balancers, the default is HTTP. For Network Load Balancers and Gateway Load Balancers, the default is TCP. The TCP protocol is not supported for health checks if the protocol of the target group is HTTP or HTTPS. It is supported for health checks only if the protocol of the target group is TCP, TLS, UDP, or TCP_UDP. The GENEVE, TLS, UDP, and TCP_UDP protocols are not supported for health checks.

With Network Load Balancers, you can't modify this setting.

healthCheckPort :: Maybe Text

The port the load balancer uses when performing health checks on targets.

healthCheckTimeoutSeconds :: Maybe Natural

HTTP/HTTPS health checks

The amount of time, in seconds, during which no response means a failed health check.

With Network Load Balancers, you can't modify this setting.

healthCheckPath :: Maybe Text

HTTP/HTTPS health checks

The destination for health checks on the targets.

HTTP1 or HTTP2 protocol version

The ping path. The default is /.

GRPC protocol version

The path of a custom health check method with the format /package.service/method. The default is /Amazon Web Services.ALB/healthcheck.

matcher :: Maybe Matcher

HTTP/HTTPS health checks

The HTTP or gRPC codes to use when checking for a successful response from a target.

With Network Load Balancers, you can't modify this setting.

healthyThresholdCount :: Maybe Natural

The number of consecutive health checks successes required before considering an unhealthy target healthy.

healthCheckIntervalSeconds :: Maybe Natural

The approximate amount of time, in seconds, between health checks of an individual target. For TCP health checks, the supported values are 10 or 30 seconds.

With Network Load Balancers, you can't modify this setting.

unhealthyThresholdCount :: Maybe Natural

The number of consecutive health check failures required before considering the target unhealthy. For target groups with a protocol of TCP or TLS, this value must be the same as the healthy threshold count.

targetGroupArn :: Text

The Amazon Resource Name (ARN) of the target group.

newModifyTargetGroup

Arguments

:: Text
$sel:targetGroupArn:ModifyTargetGroup'

-> ModifyTargetGroup

Create a value of ModifyTargetGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:healthCheckEnabled:ModifyTargetGroup', modifyTargetGroup_healthCheckEnabled - Indicates whether health checks are enabled.

$sel:healthCheckProtocol:ModifyTargetGroup', modifyTargetGroup_healthCheckProtocol - The protocol the load balancer uses when performing health checks on targets. For Application Load Balancers, the default is HTTP. For Network Load Balancers and Gateway Load Balancers, the default is TCP. The TCP protocol is not supported for health checks if the protocol of the target group is HTTP or HTTPS. It is supported for health checks only if the protocol of the target group is TCP, TLS, UDP, or TCP_UDP. The GENEVE, TLS, UDP, and TCP_UDP protocols are not supported for health checks.

With Network Load Balancers, you can't modify this setting.

$sel:healthCheckPort:ModifyTargetGroup', modifyTargetGroup_healthCheckPort - The port the load balancer uses when performing health checks on targets.

$sel:healthCheckTimeoutSeconds:ModifyTargetGroup', modifyTargetGroup_healthCheckTimeoutSeconds - [HTTP/HTTPS health checks] The amount of time, in seconds, during which no response means a failed health check.

With Network Load Balancers, you can't modify this setting.

$sel:healthCheckPath:ModifyTargetGroup', modifyTargetGroup_healthCheckPath - [HTTP/HTTPS health checks] The destination for health checks on the targets.

HTTP1 or HTTP2 protocol version

The ping path. The default is /.

GRPC protocol version

The path of a custom health check method with the format /package.service/method. The default is /Amazon Web Services.ALB/healthcheck.

$sel:matcher:ModifyTargetGroup', modifyTargetGroup_matcher - [HTTP/HTTPS health checks] The HTTP or gRPC codes to use when checking for a successful response from a target.

With Network Load Balancers, you can't modify this setting.

$sel:healthyThresholdCount:ModifyTargetGroup', modifyTargetGroup_healthyThresholdCount - The number of consecutive health checks successes required before considering an unhealthy target healthy.

$sel:healthCheckIntervalSeconds:ModifyTargetGroup', modifyTargetGroup_healthCheckIntervalSeconds - The approximate amount of time, in seconds, between health checks of an individual target. For TCP health checks, the supported values are 10 or 30 seconds.

With Network Load Balancers, you can't modify this setting.

$sel:unhealthyThresholdCount:ModifyTargetGroup', modifyTargetGroup_unhealthyThresholdCount - The number of consecutive health check failures required before considering the target unhealthy. For target groups with a protocol of TCP or TLS, this value must be the same as the healthy threshold count.

$sel:targetGroupArn:ModifyTargetGroup', modifyTargetGroup_targetGroupArn - The Amazon Resource Name (ARN) of the target group.

Request Lenses

modifyTargetGroup_healthCheckEnabled :: Lens' ModifyTargetGroup (Maybe Bool)

Indicates whether health checks are enabled.

modifyTargetGroup_healthCheckProtocol :: Lens' ModifyTargetGroup (Maybe ProtocolEnum)

The protocol the load balancer uses when performing health checks on targets. For Application Load Balancers, the default is HTTP. For Network Load Balancers and Gateway Load Balancers, the default is TCP. The TCP protocol is not supported for health checks if the protocol of the target group is HTTP or HTTPS. It is supported for health checks only if the protocol of the target group is TCP, TLS, UDP, or TCP_UDP. The GENEVE, TLS, UDP, and TCP_UDP protocols are not supported for health checks.

With Network Load Balancers, you can't modify this setting.

modifyTargetGroup_healthCheckPort :: Lens' ModifyTargetGroup (Maybe Text)

The port the load balancer uses when performing health checks on targets.

modifyTargetGroup_healthCheckTimeoutSeconds :: Lens' ModifyTargetGroup (Maybe Natural)

HTTP/HTTPS health checks

The amount of time, in seconds, during which no response means a failed health check.

With Network Load Balancers, you can't modify this setting.

modifyTargetGroup_healthCheckPath :: Lens' ModifyTargetGroup (Maybe Text)

HTTP/HTTPS health checks

The destination for health checks on the targets.

HTTP1 or HTTP2 protocol version

The ping path. The default is /.

GRPC protocol version

The path of a custom health check method with the format /package.service/method. The default is /Amazon Web Services.ALB/healthcheck.

modifyTargetGroup_matcher :: Lens' ModifyTargetGroup (Maybe Matcher)

HTTP/HTTPS health checks

The HTTP or gRPC codes to use when checking for a successful response from a target.

With Network Load Balancers, you can't modify this setting.

modifyTargetGroup_healthyThresholdCount :: Lens' ModifyTargetGroup (Maybe Natural)

The number of consecutive health checks successes required before considering an unhealthy target healthy.

modifyTargetGroup_healthCheckIntervalSeconds :: Lens' ModifyTargetGroup (Maybe Natural)

The approximate amount of time, in seconds, between health checks of an individual target. For TCP health checks, the supported values are 10 or 30 seconds.

With Network Load Balancers, you can't modify this setting.

modifyTargetGroup_unhealthyThresholdCount :: Lens' ModifyTargetGroup (Maybe Natural)

The number of consecutive health check failures required before considering the target unhealthy. For target groups with a protocol of TCP or TLS, this value must be the same as the healthy threshold count.

modifyTargetGroup_targetGroupArn :: Lens' ModifyTargetGroup Text

The Amazon Resource Name (ARN) of the target group.

Destructuring the Response

data ModifyTargetGroupResponse

See: newModifyTargetGroupResponse smart constructor.

Constructors

ModifyTargetGroupResponse'

Fields

targetGroups :: Maybe [TargetGroup]

Information about the modified target group.

httpStatus :: Int

The response's http status code.

newModifyTargetGroupResponse

Arguments

:: Int
$sel:httpStatus:ModifyTargetGroupResponse'

-> ModifyTargetGroupResponse

Create a value of ModifyTargetGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetGroups:ModifyTargetGroupResponse', modifyTargetGroupResponse_targetGroups - Information about the modified target group.

$sel:httpStatus:ModifyTargetGroupResponse', modifyTargetGroupResponse_httpStatus - The response's http status code.

Response Lenses

modifyTargetGroupResponse_targetGroups :: Lens' ModifyTargetGroupResponse (Maybe [TargetGroup])

Information about the modified target group.

modifyTargetGroupResponse_httpStatus :: Lens' ModifyTargetGroupResponse Int

The response's http status code.