Network.AWS.ELBv2.ModifyTargetGroupAttributes

Description

Modifies the specified attributes of the specified target group.

Creating a Request

data ModifyTargetGroupAttributes

See: newModifyTargetGroupAttributes smart constructor.

Constructors

ModifyTargetGroupAttributes'

Fields

targetGroupArn :: Text

The Amazon Resource Name (ARN) of the target group.

attributes :: [TargetGroupAttribute]

The attributes.

newModifyTargetGroupAttributes

Arguments

:: Text
$sel:targetGroupArn:ModifyTargetGroupAttributes'

-> ModifyTargetGroupAttributes

Create a value of ModifyTargetGroupAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetGroupArn:ModifyTargetGroupAttributes', modifyTargetGroupAttributes_targetGroupArn - The Amazon Resource Name (ARN) of the target group.

$sel:attributes:ModifyTargetGroupAttributes', modifyTargetGroupAttributes_attributes - The attributes.

Request Lenses

modifyTargetGroupAttributes_targetGroupArn :: Lens' ModifyTargetGroupAttributes Text

The Amazon Resource Name (ARN) of the target group.

modifyTargetGroupAttributes_attributes :: Lens' ModifyTargetGroupAttributes [TargetGroupAttribute]

The attributes.

Destructuring the Response

data ModifyTargetGroupAttributesResponse

See: newModifyTargetGroupAttributesResponse smart constructor.

Constructors

ModifyTargetGroupAttributesResponse'

Fields

attributes :: Maybe [TargetGroupAttribute]

Information about the attributes.

httpStatus :: Int

The response's http status code.

newModifyTargetGroupAttributesResponse

Arguments

:: Int
$sel:httpStatus:ModifyTargetGroupAttributesResponse'

-> ModifyTargetGroupAttributesResponse

Create a value of ModifyTargetGroupAttributesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributes:ModifyTargetGroupAttributes', modifyTargetGroupAttributesResponse_attributes - Information about the attributes.

$sel:httpStatus:ModifyTargetGroupAttributesResponse', modifyTargetGroupAttributesResponse_httpStatus - The response's http status code.

Response Lenses

modifyTargetGroupAttributesResponse_attributes :: Lens' ModifyTargetGroupAttributesResponse (Maybe [TargetGroupAttribute])

Information about the attributes.

modifyTargetGroupAttributesResponse_httpStatus :: Lens' ModifyTargetGroupAttributesResponse Int

The response's http status code.