Network.AWS.ELBv2.Types.Action

Description

Documentation

data Action

Information about an action.

Each rule must include exactly one of the following types of actions: forward, fixed-response, or redirect, and it must be the last action to be performed.

See: newAction smart constructor.

Constructors

Action'

Fields

authenticateOidcConfig :: Maybe AuthenticateOidcActionConfig

HTTPS listeners

Information about an identity provider that is compliant with OpenID Connect (OIDC). Specify only when Type is authenticate-oidc.

targetGroupArn :: Maybe Text

The Amazon Resource Name (ARN) of the target group. Specify only when Type is forward and you want to route to a single target group. To route to one or more target groups, use ForwardConfig instead.

authenticateCognitoConfig :: Maybe AuthenticateCognitoActionConfig

HTTPS listeners

Information for using Amazon Cognito to authenticate users. Specify only when Type is authenticate-cognito.

fixedResponseConfig :: Maybe FixedResponseActionConfig

Application Load Balancer

Information for creating an action that returns a custom HTTP response. Specify only when Type is fixed-response.

forwardConfig :: Maybe ForwardActionConfig

Information for creating an action that distributes requests among one or more target groups. For Network Load Balancers, you can specify a single target group. Specify only when Type is forward. If you specify both ForwardConfig and TargetGroupArn, you can specify only one target group using ForwardConfig and it must be the same target group specified in TargetGroupArn.

order :: Maybe Natural

The order for the action. This value is required for rules with multiple actions. The action with the lowest value for order is performed first.

redirectConfig :: Maybe RedirectActionConfig

Application Load Balancer

Information for creating a redirect action. Specify only when Type is redirect.

type' :: ActionTypeEnum

The type of action.

newAction

Arguments

:: ActionTypeEnum
$sel:type':Action'

-> Action

Create a value of Action with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:authenticateOidcConfig:Action', action_authenticateOidcConfig - [HTTPS listeners] Information about an identity provider that is compliant with OpenID Connect (OIDC). Specify only when Type is authenticate-oidc.

$sel:targetGroupArn:Action', action_targetGroupArn - The Amazon Resource Name (ARN) of the target group. Specify only when Type is forward and you want to route to a single target group. To route to one or more target groups, use ForwardConfig instead.

$sel:authenticateCognitoConfig:Action', action_authenticateCognitoConfig - [HTTPS listeners] Information for using Amazon Cognito to authenticate users. Specify only when Type is authenticate-cognito.

$sel:fixedResponseConfig:Action', action_fixedResponseConfig - [Application Load Balancer] Information for creating an action that returns a custom HTTP response. Specify only when Type is fixed-response.

$sel:forwardConfig:Action', action_forwardConfig - Information for creating an action that distributes requests among one or more target groups. For Network Load Balancers, you can specify a single target group. Specify only when Type is forward. If you specify both ForwardConfig and TargetGroupArn, you can specify only one target group using ForwardConfig and it must be the same target group specified in TargetGroupArn.

$sel:order:Action', action_order - The order for the action. This value is required for rules with multiple actions. The action with the lowest value for order is performed first.

$sel:redirectConfig:Action', action_redirectConfig - [Application Load Balancer] Information for creating a redirect action. Specify only when Type is redirect.

$sel:type':Action', action_type - The type of action.

action_authenticateOidcConfig :: Lens' Action (Maybe AuthenticateOidcActionConfig)

HTTPS listeners

Information about an identity provider that is compliant with OpenID Connect (OIDC). Specify only when Type is authenticate-oidc.

action_targetGroupArn :: Lens' Action (Maybe Text)

The Amazon Resource Name (ARN) of the target group. Specify only when Type is forward and you want to route to a single target group. To route to one or more target groups, use ForwardConfig instead.

action_authenticateCognitoConfig :: Lens' Action (Maybe AuthenticateCognitoActionConfig)

HTTPS listeners

Information for using Amazon Cognito to authenticate users. Specify only when Type is authenticate-cognito.

action_fixedResponseConfig :: Lens' Action (Maybe FixedResponseActionConfig)

Application Load Balancer

Information for creating an action that returns a custom HTTP response. Specify only when Type is fixed-response.

action_forwardConfig :: Lens' Action (Maybe ForwardActionConfig)

Information for creating an action that distributes requests among one or more target groups. For Network Load Balancers, you can specify a single target group. Specify only when Type is forward. If you specify both ForwardConfig and TargetGroupArn, you can specify only one target group using ForwardConfig and it must be the same target group specified in TargetGroupArn.

action_order :: Lens' Action (Maybe Natural)

The order for the action. This value is required for rules with multiple actions. The action with the lowest value for order is performed first.

action_redirectConfig :: Lens' Action (Maybe RedirectActionConfig)

Application Load Balancer

Information for creating a redirect action. Specify only when Type is redirect.

action_type :: Lens' Action ActionTypeEnum

The type of action.