Network.AWS.ELBv2.Types.ForwardActionConfig

Description

Documentation

data ForwardActionConfig

Information about a forward action.

See: newForwardActionConfig smart constructor.

Constructors

ForwardActionConfig'

Fields

targetGroupStickinessConfig :: Maybe TargetGroupStickinessConfig

The target group stickiness for the rule.

targetGroups :: Maybe [TargetGroupTuple]

One or more target groups. For Network Load Balancers, you can specify a single target group.

newForwardActionConfig :: ForwardActionConfig

Create a value of ForwardActionConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetGroupStickinessConfig:ForwardActionConfig', forwardActionConfig_targetGroupStickinessConfig - The target group stickiness for the rule.

$sel:targetGroups:ForwardActionConfig', forwardActionConfig_targetGroups - One or more target groups. For Network Load Balancers, you can specify a single target group.

forwardActionConfig_targetGroupStickinessConfig :: Lens' ForwardActionConfig (Maybe TargetGroupStickinessConfig)

The target group stickiness for the rule.

forwardActionConfig_targetGroups :: Lens' ForwardActionConfig (Maybe [TargetGroupTuple])

One or more target groups. For Network Load Balancers, you can specify a single target group.