Network.AWS.ELBv2.Types.Limit

Description

Documentation

data Limit

Information about an Elastic Load Balancing resource limit for your Amazon Web Services account.

See: newLimit smart constructor.

Constructors

Limit'

Fields

name :: Maybe Text

The name of the limit. The possible values are:

application-load-balancers

condition-values-per-alb-rule

condition-wildcards-per-alb-rule

gateway-load-balancers

gateway-load-balancers-per-vpc

geneve-target-groups

listeners-per-application-load-balancer

listeners-per-network-load-balancer

network-load-balancers

rules-per-application-load-balancer

target-groups

target-groups-per-action-on-application-load-balancer

target-groups-per-action-on-network-load-balancer

target-groups-per-application-load-balancer

targets-per-application-load-balancer

targets-per-availability-zone-per-gateway-load-balancer

targets-per-availability-zone-per-network-load-balancer

targets-per-network-load-balancer

max :: Maybe Text

The maximum value of the limit.

newLimit :: Limit

Create a value of Limit with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:Limit', limit_name - The name of the limit. The possible values are:

application-load-balancers

condition-values-per-alb-rule

condition-wildcards-per-alb-rule

gateway-load-balancers

gateway-load-balancers-per-vpc

geneve-target-groups

listeners-per-application-load-balancer

listeners-per-network-load-balancer

network-load-balancers

rules-per-application-load-balancer

target-groups

target-groups-per-action-on-application-load-balancer

target-groups-per-action-on-network-load-balancer

target-groups-per-application-load-balancer

targets-per-application-load-balancer

targets-per-availability-zone-per-gateway-load-balancer

targets-per-availability-zone-per-network-load-balancer

targets-per-network-load-balancer

$sel:max:Limit', limit_max - The maximum value of the limit.

limit_name :: Lens' Limit (Maybe Text)

The name of the limit. The possible values are:

application-load-balancers

condition-values-per-alb-rule

condition-wildcards-per-alb-rule

gateway-load-balancers

gateway-load-balancers-per-vpc

geneve-target-groups

listeners-per-application-load-balancer

listeners-per-network-load-balancer

network-load-balancers

rules-per-application-load-balancer

target-groups

target-groups-per-action-on-application-load-balancer

target-groups-per-action-on-network-load-balancer

target-groups-per-application-load-balancer

targets-per-application-load-balancer

targets-per-availability-zone-per-gateway-load-balancer

targets-per-availability-zone-per-network-load-balancer

targets-per-network-load-balancer

limit_max :: Lens' Limit (Maybe Text)

The maximum value of the limit.