Network.AWS.ELBv2.Types.LoadBalancer

Description

Documentation

data LoadBalancer

Information about a load balancer.

See: newLoadBalancer smart constructor.

Constructors

LoadBalancer'

Fields

ipAddressType :: Maybe IpAddressType

The type of IP addresses used by the subnets for your load balancer. The possible values are ipv4 (for IPv4 addresses) and dualstack (for IPv4 and IPv6 addresses).

loadBalancerArn :: Maybe Text

The Amazon Resource Name (ARN) of the load balancer.

availabilityZones :: Maybe [AvailabilityZone]

The subnets for the load balancer.

customerOwnedIpv4Pool :: Maybe Text

Application Load Balancers on Outposts

The ID of the customer-owned address pool.

scheme :: Maybe LoadBalancerSchemeEnum

The nodes of an Internet-facing load balancer have public IP addresses. The DNS name of an Internet-facing load balancer is publicly resolvable to the public IP addresses of the nodes. Therefore, Internet-facing load balancers can route requests from clients over the internet.

The nodes of an internal load balancer have only private IP addresses. The DNS name of an internal load balancer is publicly resolvable to the private IP addresses of the nodes. Therefore, internal load balancers can route requests only from clients with access to the VPC for the load balancer.

createdTime :: Maybe ISO8601

The date and time the load balancer was created.

securityGroups :: Maybe [Text]

The IDs of the security groups for the load balancer.

state :: Maybe LoadBalancerState

The state of the load balancer.

dNSName :: Maybe Text

The public DNS name of the load balancer.

canonicalHostedZoneId :: Maybe Text

The ID of the Amazon Route 53 hosted zone associated with the load balancer.

vpcId :: Maybe Text

The ID of the VPC for the load balancer.

type' :: Maybe LoadBalancerTypeEnum

The type of load balancer.

loadBalancerName :: Maybe Text

The name of the load balancer.

newLoadBalancer :: LoadBalancer

Create a value of LoadBalancer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ipAddressType:LoadBalancer', loadBalancer_ipAddressType - The type of IP addresses used by the subnets for your load balancer. The possible values are ipv4 (for IPv4 addresses) and dualstack (for IPv4 and IPv6 addresses).

$sel:loadBalancerArn:LoadBalancer', loadBalancer_loadBalancerArn - The Amazon Resource Name (ARN) of the load balancer.

$sel:availabilityZones:LoadBalancer', loadBalancer_availabilityZones - The subnets for the load balancer.

$sel:customerOwnedIpv4Pool:LoadBalancer', loadBalancer_customerOwnedIpv4Pool - [Application Load Balancers on Outposts] The ID of the customer-owned address pool.

$sel:scheme:LoadBalancer', loadBalancer_scheme - The nodes of an Internet-facing load balancer have public IP addresses. The DNS name of an Internet-facing load balancer is publicly resolvable to the public IP addresses of the nodes. Therefore, Internet-facing load balancers can route requests from clients over the internet.

The nodes of an internal load balancer have only private IP addresses. The DNS name of an internal load balancer is publicly resolvable to the private IP addresses of the nodes. Therefore, internal load balancers can route requests only from clients with access to the VPC for the load balancer.

$sel:createdTime:LoadBalancer', loadBalancer_createdTime - The date and time the load balancer was created.

$sel:securityGroups:LoadBalancer', loadBalancer_securityGroups - The IDs of the security groups for the load balancer.

$sel:state:LoadBalancer', loadBalancer_state - The state of the load balancer.

$sel:dNSName:LoadBalancer', loadBalancer_dNSName - The public DNS name of the load balancer.

$sel:canonicalHostedZoneId:LoadBalancer', loadBalancer_canonicalHostedZoneId - The ID of the Amazon Route 53 hosted zone associated with the load balancer.

$sel:vpcId:LoadBalancer', loadBalancer_vpcId - The ID of the VPC for the load balancer.

$sel:type':LoadBalancer', loadBalancer_type - The type of load balancer.

$sel:loadBalancerName:LoadBalancer', loadBalancer_loadBalancerName - The name of the load balancer.

loadBalancer_ipAddressType :: Lens' LoadBalancer (Maybe IpAddressType)

The type of IP addresses used by the subnets for your load balancer. The possible values are ipv4 (for IPv4 addresses) and dualstack (for IPv4 and IPv6 addresses).

loadBalancer_loadBalancerArn :: Lens' LoadBalancer (Maybe Text)

The Amazon Resource Name (ARN) of the load balancer.

loadBalancer_availabilityZones :: Lens' LoadBalancer (Maybe [AvailabilityZone])

The subnets for the load balancer.

loadBalancer_customerOwnedIpv4Pool :: Lens' LoadBalancer (Maybe Text)

Application Load Balancers on Outposts

The ID of the customer-owned address pool.

loadBalancer_scheme :: Lens' LoadBalancer (Maybe LoadBalancerSchemeEnum)

The nodes of an Internet-facing load balancer have public IP addresses. The DNS name of an Internet-facing load balancer is publicly resolvable to the public IP addresses of the nodes. Therefore, Internet-facing load balancers can route requests from clients over the internet.

The nodes of an internal load balancer have only private IP addresses. The DNS name of an internal load balancer is publicly resolvable to the private IP addresses of the nodes. Therefore, internal load balancers can route requests only from clients with access to the VPC for the load balancer.

loadBalancer_createdTime :: Lens' LoadBalancer (Maybe UTCTime)

The date and time the load balancer was created.

loadBalancer_securityGroups :: Lens' LoadBalancer (Maybe [Text])

The IDs of the security groups for the load balancer.

loadBalancer_state :: Lens' LoadBalancer (Maybe LoadBalancerState)

The state of the load balancer.

loadBalancer_dNSName :: Lens' LoadBalancer (Maybe Text)

The public DNS name of the load balancer.

loadBalancer_canonicalHostedZoneId :: Lens' LoadBalancer (Maybe Text)

The ID of the Amazon Route 53 hosted zone associated with the load balancer.

loadBalancer_vpcId :: Lens' LoadBalancer (Maybe Text)

The ID of the VPC for the load balancer.

loadBalancer_type :: Lens' LoadBalancer (Maybe LoadBalancerTypeEnum)

The type of load balancer.

loadBalancer_loadBalancerName :: Lens' LoadBalancer (Maybe Text)

The name of the load balancer.