Network.AWS.ELBv2.Types.Matcher

Description

Documentation

data Matcher

The codes to use when checking for a successful response from a target. If the protocol version is gRPC, these are gRPC codes. Otherwise, these are HTTP codes.

See: newMatcher smart constructor.

Constructors

Matcher'

Fields

httpCode :: Maybe Text

For Application Load Balancers, you can specify values between 200 and 499, and the default value is 200. You can specify multiple values (for example, "200,202") or a range of values (for example, "200-299").

For Network Load Balancers and Gateway Load Balancers, this must be "200–399".

Note that when using shorthand syntax, some values such as commas need to be escaped.

grpcCode :: Maybe Text

You can specify values between 0 and 99. You can specify multiple values (for example, "0,1") or a range of values (for example, "0-5"). The default value is 12.

newMatcher :: Matcher

Create a value of Matcher with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpCode:Matcher', matcher_httpCode - For Application Load Balancers, you can specify values between 200 and 499, and the default value is 200. You can specify multiple values (for example, "200,202") or a range of values (for example, "200-299").

For Network Load Balancers and Gateway Load Balancers, this must be "200–399".

Note that when using shorthand syntax, some values such as commas need to be escaped.

$sel:grpcCode:Matcher', matcher_grpcCode - You can specify values between 0 and 99. You can specify multiple values (for example, "0,1") or a range of values (for example, "0-5"). The default value is 12.

matcher_httpCode :: Lens' Matcher (Maybe Text)

For Application Load Balancers, you can specify values between 200 and 499, and the default value is 200. You can specify multiple values (for example, "200,202") or a range of values (for example, "200-299").

For Network Load Balancers and Gateway Load Balancers, this must be "200–399".

Note that when using shorthand syntax, some values such as commas need to be escaped.

matcher_grpcCode :: Lens' Matcher (Maybe Text)

You can specify values between 0 and 99. You can specify multiple values (for example, "0,1") or a range of values (for example, "0-5"). The default value is 12.