Network.AWS.ELBv2.Types.RedirectActionConfig

Description

Documentation

data RedirectActionConfig

Information about a redirect action.

A URI consists of the following components: protocol://hostname:port/path?query. You must modify at least one of the following components to avoid a redirect loop: protocol, hostname, port, or path. Any components that you do not modify retain their original values.

You can reuse URI components using the following reserved keywords:

#{protocol}

#{host}

#{port}

#{path} (the leading "/" is removed)

#{query}

For example, you can change the path to "/new/#{path}", the hostname to "example.#{host}", or the query to "#{query}&value=xyz".

See: newRedirectActionConfig smart constructor.

Constructors

RedirectActionConfig'

Fields

query :: Maybe Text

The query parameters, URL-encoded when necessary, but not percent-encoded. Do not include the leading "?", as it is automatically added. You can specify any of the reserved keywords.

port :: Maybe Text

The port. You can specify a value from 1 to 65535 or #{port}.

protocol :: Maybe Text

The protocol. You can specify HTTP, HTTPS, or #{protocol}. You can redirect HTTP to HTTP, HTTP to HTTPS, and HTTPS to HTTPS. You cannot redirect HTTPS to HTTP.

host :: Maybe Text

The hostname. This component is not percent-encoded. The hostname can contain #{host}.

path :: Maybe Text

The absolute path, starting with the leading "/". This component is not percent-encoded. The path can contain #{host}, #{path}, and #{port}.

statusCode :: RedirectActionStatusCodeEnum

The HTTP redirect code. The redirect is either permanent (HTTP 301) or temporary (HTTP 302).

newRedirectActionConfig

Arguments

:: RedirectActionStatusCodeEnum
$sel:statusCode:RedirectActionConfig'

-> RedirectActionConfig

Create a value of RedirectActionConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:query:RedirectActionConfig', redirectActionConfig_query - The query parameters, URL-encoded when necessary, but not percent-encoded. Do not include the leading "?", as it is automatically added. You can specify any of the reserved keywords.

$sel:port:RedirectActionConfig', redirectActionConfig_port - The port. You can specify a value from 1 to 65535 or #{port}.

$sel:protocol:RedirectActionConfig', redirectActionConfig_protocol - The protocol. You can specify HTTP, HTTPS, or #{protocol}. You can redirect HTTP to HTTP, HTTP to HTTPS, and HTTPS to HTTPS. You cannot redirect HTTPS to HTTP.

$sel:host:RedirectActionConfig', redirectActionConfig_host - The hostname. This component is not percent-encoded. The hostname can contain #{host}.

$sel:path:RedirectActionConfig', redirectActionConfig_path - The absolute path, starting with the leading "/". This component is not percent-encoded. The path can contain #{host}, #{path}, and #{port}.

$sel:statusCode:RedirectActionConfig', redirectActionConfig_statusCode - The HTTP redirect code. The redirect is either permanent (HTTP 301) or temporary (HTTP 302).

redirectActionConfig_query :: Lens' RedirectActionConfig (Maybe Text)

The query parameters, URL-encoded when necessary, but not percent-encoded. Do not include the leading "?", as it is automatically added. You can specify any of the reserved keywords.

redirectActionConfig_port :: Lens' RedirectActionConfig (Maybe Text)

The port. You can specify a value from 1 to 65535 or #{port}.

redirectActionConfig_protocol :: Lens' RedirectActionConfig (Maybe Text)

The protocol. You can specify HTTP, HTTPS, or #{protocol}. You can redirect HTTP to HTTP, HTTP to HTTPS, and HTTPS to HTTPS. You cannot redirect HTTPS to HTTP.

redirectActionConfig_host :: Lens' RedirectActionConfig (Maybe Text)

The hostname. This component is not percent-encoded. The hostname can contain #{host}.

redirectActionConfig_path :: Lens' RedirectActionConfig (Maybe Text)

The absolute path, starting with the leading "/". This component is not percent-encoded. The path can contain #{host}, #{path}, and #{port}.

redirectActionConfig_statusCode :: Lens' RedirectActionConfig RedirectActionStatusCodeEnum

The HTTP redirect code. The redirect is either permanent (HTTP 301) or temporary (HTTP 302).