Network.AWS.ELBv2.Types.TargetGroup

Description

Documentation

data TargetGroup

Information about a target group.

See: newTargetGroup smart constructor.

Constructors

TargetGroup'

Fields

protocolVersion :: Maybe Text

HTTP/HTTPS protocol

The protocol version. The possible values are GRPC, HTTP1, and HTTP2.

matcher :: Maybe Matcher

The HTTP or gRPC codes to use when checking for a successful response from a target.

healthCheckPath :: Maybe Text

The destination for health checks on the targets.

healthCheckEnabled :: Maybe Bool

Indicates whether health checks are enabled.

unhealthyThresholdCount :: Maybe Natural

The number of consecutive health check failures required before considering the target unhealthy.

vpcId :: Maybe Text

The ID of the VPC for the targets.

targetGroupArn :: Maybe Text

The Amazon Resource Name (ARN) of the target group.

protocol :: Maybe ProtocolEnum

The protocol to use for routing traffic to the targets.

healthCheckIntervalSeconds :: Maybe Natural

The approximate amount of time, in seconds, between health checks of an individual target.

targetType :: Maybe TargetTypeEnum

The type of target that you must specify when registering targets with this target group. The possible values are instance (register targets by instance ID), ip (register targets by IP address), lambda (register a single Lambda function as a target), or alb (register a single Application Load Balancer as a target).

healthyThresholdCount :: Maybe Natural

The number of consecutive health checks successes required before considering an unhealthy target healthy.

healthCheckProtocol :: Maybe ProtocolEnum

The protocol to use to connect with the target. The GENEVE, TLS, UDP, and TCP_UDP protocols are not supported for health checks.

loadBalancerArns :: Maybe [Text]

The Amazon Resource Names (ARN) of the load balancers that route traffic to this target group.

ipAddressType :: Maybe TargetGroupIpAddressTypeEnum

The type of IP address used for this target group. The possible values are ipv4 and ipv6. This is an optional parameter. If not specified, the IP address type defaults to ipv4.

healthCheckTimeoutSeconds :: Maybe Natural

The amount of time, in seconds, during which no response means a failed health check.

healthCheckPort :: Maybe Text

The port to use to connect with the target.

targetGroupName :: Maybe Text

The name of the target group.

port :: Maybe Natural

The port on which the targets are listening. Not used if the target is a Lambda function.

newTargetGroup :: TargetGroup

Create a value of TargetGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:protocolVersion:TargetGroup', targetGroup_protocolVersion - [HTTP/HTTPS protocol] The protocol version. The possible values are GRPC, HTTP1, and HTTP2.

$sel:matcher:TargetGroup', targetGroup_matcher - The HTTP or gRPC codes to use when checking for a successful response from a target.

$sel:healthCheckPath:TargetGroup', targetGroup_healthCheckPath - The destination for health checks on the targets.

$sel:healthCheckEnabled:TargetGroup', targetGroup_healthCheckEnabled - Indicates whether health checks are enabled.

$sel:unhealthyThresholdCount:TargetGroup', targetGroup_unhealthyThresholdCount - The number of consecutive health check failures required before considering the target unhealthy.

$sel:vpcId:TargetGroup', targetGroup_vpcId - The ID of the VPC for the targets.

$sel:targetGroupArn:TargetGroup', targetGroup_targetGroupArn - The Amazon Resource Name (ARN) of the target group.

$sel:protocol:TargetGroup', targetGroup_protocol - The protocol to use for routing traffic to the targets.

$sel:healthCheckIntervalSeconds:TargetGroup', targetGroup_healthCheckIntervalSeconds - The approximate amount of time, in seconds, between health checks of an individual target.

$sel:targetType:TargetGroup', targetGroup_targetType - The type of target that you must specify when registering targets with this target group. The possible values are instance (register targets by instance ID), ip (register targets by IP address), lambda (register a single Lambda function as a target), or alb (register a single Application Load Balancer as a target).

$sel:healthyThresholdCount:TargetGroup', targetGroup_healthyThresholdCount - The number of consecutive health checks successes required before considering an unhealthy target healthy.

$sel:healthCheckProtocol:TargetGroup', targetGroup_healthCheckProtocol - The protocol to use to connect with the target. The GENEVE, TLS, UDP, and TCP_UDP protocols are not supported for health checks.

$sel:loadBalancerArns:TargetGroup', targetGroup_loadBalancerArns - The Amazon Resource Names (ARN) of the load balancers that route traffic to this target group.

$sel:ipAddressType:TargetGroup', targetGroup_ipAddressType - The type of IP address used for this target group. The possible values are ipv4 and ipv6. This is an optional parameter. If not specified, the IP address type defaults to ipv4.

$sel:healthCheckTimeoutSeconds:TargetGroup', targetGroup_healthCheckTimeoutSeconds - The amount of time, in seconds, during which no response means a failed health check.

$sel:healthCheckPort:TargetGroup', targetGroup_healthCheckPort - The port to use to connect with the target.

$sel:targetGroupName:TargetGroup', targetGroup_targetGroupName - The name of the target group.

$sel:port:TargetGroup', targetGroup_port - The port on which the targets are listening. Not used if the target is a Lambda function.

targetGroup_protocolVersion :: Lens' TargetGroup (Maybe Text)

HTTP/HTTPS protocol

The protocol version. The possible values are GRPC, HTTP1, and HTTP2.

targetGroup_matcher :: Lens' TargetGroup (Maybe Matcher)

The HTTP or gRPC codes to use when checking for a successful response from a target.

targetGroup_healthCheckPath :: Lens' TargetGroup (Maybe Text)

The destination for health checks on the targets.

targetGroup_healthCheckEnabled :: Lens' TargetGroup (Maybe Bool)

Indicates whether health checks are enabled.

targetGroup_unhealthyThresholdCount :: Lens' TargetGroup (Maybe Natural)

The number of consecutive health check failures required before considering the target unhealthy.

targetGroup_vpcId :: Lens' TargetGroup (Maybe Text)

The ID of the VPC for the targets.

targetGroup_targetGroupArn :: Lens' TargetGroup (Maybe Text)

The Amazon Resource Name (ARN) of the target group.

targetGroup_protocol :: Lens' TargetGroup (Maybe ProtocolEnum)

The protocol to use for routing traffic to the targets.

targetGroup_healthCheckIntervalSeconds :: Lens' TargetGroup (Maybe Natural)

The approximate amount of time, in seconds, between health checks of an individual target.

targetGroup_targetType :: Lens' TargetGroup (Maybe TargetTypeEnum)

The type of target that you must specify when registering targets with this target group. The possible values are instance (register targets by instance ID), ip (register targets by IP address), lambda (register a single Lambda function as a target), or alb (register a single Application Load Balancer as a target).

targetGroup_healthyThresholdCount :: Lens' TargetGroup (Maybe Natural)

The number of consecutive health checks successes required before considering an unhealthy target healthy.

targetGroup_healthCheckProtocol :: Lens' TargetGroup (Maybe ProtocolEnum)

The protocol to use to connect with the target. The GENEVE, TLS, UDP, and TCP_UDP protocols are not supported for health checks.

targetGroup_loadBalancerArns :: Lens' TargetGroup (Maybe [Text])

The Amazon Resource Names (ARN) of the load balancers that route traffic to this target group.

targetGroup_ipAddressType :: Lens' TargetGroup (Maybe TargetGroupIpAddressTypeEnum)

The type of IP address used for this target group. The possible values are ipv4 and ipv6. This is an optional parameter. If not specified, the IP address type defaults to ipv4.

targetGroup_healthCheckTimeoutSeconds :: Lens' TargetGroup (Maybe Natural)

The amount of time, in seconds, during which no response means a failed health check.

targetGroup_healthCheckPort :: Lens' TargetGroup (Maybe Text)

The port to use to connect with the target.

targetGroup_targetGroupName :: Lens' TargetGroup (Maybe Text)

The name of the target group.

targetGroup_port :: Lens' TargetGroup (Maybe Natural)

The port on which the targets are listening. Not used if the target is a Lambda function.