Network.AWS.ELBv2.Types.TargetHealthDescription

Description

Documentation

data TargetHealthDescription

Information about the health of a target.

See: newTargetHealthDescription smart constructor.

Constructors

TargetHealthDescription'

Fields

healthCheckPort :: Maybe Text

The port to use to connect with the target.

target :: Maybe TargetDescription

The description of the target.

targetHealth :: Maybe TargetHealth

The health information for the target.

newTargetHealthDescription :: TargetHealthDescription

Create a value of TargetHealthDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:healthCheckPort:TargetHealthDescription', targetHealthDescription_healthCheckPort - The port to use to connect with the target.

$sel:target:TargetHealthDescription', targetHealthDescription_target - The description of the target.

$sel:targetHealth:TargetHealthDescription', targetHealthDescription_targetHealth - The health information for the target.

targetHealthDescription_healthCheckPort :: Lens' TargetHealthDescription (Maybe Text)

The port to use to connect with the target.

targetHealthDescription_target :: Lens' TargetHealthDescription (Maybe TargetDescription)

The description of the target.

targetHealthDescription_targetHealth :: Lens' TargetHealthDescription (Maybe TargetHealth)

The health information for the target.