Network.AWS.EMR.AddInstanceFleet

Description

Adds an instance fleet to a running cluster.

The instance fleet configuration is available only in Amazon EMR versions 4.8.0 and later, excluding 5.0.x.

Creating a Request

data AddInstanceFleet

See: newAddInstanceFleet smart constructor.

Constructors

AddInstanceFleet'

Fields

clusterId :: Text

The unique identifier of the cluster.

instanceFleet :: InstanceFleetConfig

Specifies the configuration of the instance fleet.

newAddInstanceFleet

Arguments

:: Text
$sel:clusterId:AddInstanceFleet'

-> InstanceFleetConfig
$sel:instanceFleet:AddInstanceFleet'

-> AddInstanceFleet

Create a value of AddInstanceFleet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clusterId:AddInstanceFleet', addInstanceFleet_clusterId - The unique identifier of the cluster.

$sel:instanceFleet:AddInstanceFleet', addInstanceFleet_instanceFleet - Specifies the configuration of the instance fleet.

Request Lenses

addInstanceFleet_clusterId :: Lens' AddInstanceFleet Text

The unique identifier of the cluster.

addInstanceFleet_instanceFleet :: Lens' AddInstanceFleet InstanceFleetConfig

Specifies the configuration of the instance fleet.

Destructuring the Response

data AddInstanceFleetResponse

See: newAddInstanceFleetResponse smart constructor.

Constructors

AddInstanceFleetResponse'

Fields

clusterArn :: Maybe Text

The Amazon Resource Name of the cluster.

clusterId :: Maybe Text

The unique identifier of the cluster.

instanceFleetId :: Maybe Text

The unique identifier of the instance fleet.

httpStatus :: Int

The response's http status code.

newAddInstanceFleetResponse

Arguments

:: Int
$sel:httpStatus:AddInstanceFleetResponse'

-> AddInstanceFleetResponse

Create a value of AddInstanceFleetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clusterArn:AddInstanceFleetResponse', addInstanceFleetResponse_clusterArn - The Amazon Resource Name of the cluster.

$sel:clusterId:AddInstanceFleet', addInstanceFleetResponse_clusterId - The unique identifier of the cluster.

$sel:instanceFleetId:AddInstanceFleetResponse', addInstanceFleetResponse_instanceFleetId - The unique identifier of the instance fleet.

$sel:httpStatus:AddInstanceFleetResponse', addInstanceFleetResponse_httpStatus - The response's http status code.

Response Lenses

addInstanceFleetResponse_clusterArn :: Lens' AddInstanceFleetResponse (Maybe Text)

The Amazon Resource Name of the cluster.

addInstanceFleetResponse_clusterId :: Lens' AddInstanceFleetResponse (Maybe Text)

The unique identifier of the cluster.

addInstanceFleetResponse_instanceFleetId :: Lens' AddInstanceFleetResponse (Maybe Text)

The unique identifier of the instance fleet.

addInstanceFleetResponse_httpStatus :: Lens' AddInstanceFleetResponse Int

The response's http status code.