Network.AWS.EMR.DescribeReleaseLabel

Description

Provides EMR release label details, such as releases available the region where the API request is run, and the available applications for a specific EMR release label. Can also list EMR release versions that support a specified version of Spark.

Creating a Request

data DescribeReleaseLabel

See: newDescribeReleaseLabel smart constructor.

Constructors

DescribeReleaseLabel'

Fields

nextToken :: Maybe Text

The pagination token. Reserved for future use. Currently set to null.

maxResults :: Maybe Natural

Reserved for future use. Currently set to null.

releaseLabel :: Maybe Text

The target release label to be described.

newDescribeReleaseLabel :: DescribeReleaseLabel

Create a value of DescribeReleaseLabel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeReleaseLabel', describeReleaseLabel_nextToken - The pagination token. Reserved for future use. Currently set to null.

$sel:maxResults:DescribeReleaseLabel', describeReleaseLabel_maxResults - Reserved for future use. Currently set to null.

$sel:releaseLabel:DescribeReleaseLabel', describeReleaseLabel_releaseLabel - The target release label to be described.

Request Lenses

describeReleaseLabel_nextToken :: Lens' DescribeReleaseLabel (Maybe Text)

The pagination token. Reserved for future use. Currently set to null.

describeReleaseLabel_maxResults :: Lens' DescribeReleaseLabel (Maybe Natural)

Reserved for future use. Currently set to null.

describeReleaseLabel_releaseLabel :: Lens' DescribeReleaseLabel (Maybe Text)

The target release label to be described.

Destructuring the Response

data DescribeReleaseLabelResponse

See: newDescribeReleaseLabelResponse smart constructor.

Constructors

DescribeReleaseLabelResponse'

Fields

nextToken :: Maybe Text

The pagination token. Reserved for future use. Currently set to null.

releaseLabel :: Maybe Text

The target release label described in the response.

applications :: Maybe [SimplifiedApplication]

The list of applications available for the target release label. Name is the name of the application. Version is the concise version of the application.

httpStatus :: Int

The response's http status code.

newDescribeReleaseLabelResponse

Arguments

:: Int
$sel:httpStatus:DescribeReleaseLabelResponse'

-> DescribeReleaseLabelResponse

Create a value of DescribeReleaseLabelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeReleaseLabel', describeReleaseLabelResponse_nextToken - The pagination token. Reserved for future use. Currently set to null.

$sel:releaseLabel:DescribeReleaseLabel', describeReleaseLabelResponse_releaseLabel - The target release label described in the response.

$sel:applications:DescribeReleaseLabelResponse', describeReleaseLabelResponse_applications - The list of applications available for the target release label. Name is the name of the application. Version is the concise version of the application.

$sel:httpStatus:DescribeReleaseLabelResponse', describeReleaseLabelResponse_httpStatus - The response's http status code.

Response Lenses

describeReleaseLabelResponse_nextToken :: Lens' DescribeReleaseLabelResponse (Maybe Text)

The pagination token. Reserved for future use. Currently set to null.

describeReleaseLabelResponse_releaseLabel :: Lens' DescribeReleaseLabelResponse (Maybe Text)

The target release label described in the response.

describeReleaseLabelResponse_applications :: Lens' DescribeReleaseLabelResponse (Maybe [SimplifiedApplication])

The list of applications available for the target release label. Name is the name of the application. Version is the concise version of the application.

describeReleaseLabelResponse_httpStatus :: Lens' DescribeReleaseLabelResponse Int

The response's http status code.