Network.AWS.EMR.DescribeStudio

Description

Returns details for the specified Amazon EMR Studio including ID, Name, VPC, Studio access URL, and so on.

Creating a Request

data DescribeStudio

See: newDescribeStudio smart constructor.

Constructors

DescribeStudio'

Fields

studioId :: Text

The Amazon EMR Studio ID.

newDescribeStudio

Arguments

:: Text
$sel:studioId:DescribeStudio'

-> DescribeStudio

Create a value of DescribeStudio with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:studioId:DescribeStudio', describeStudio_studioId - The Amazon EMR Studio ID.

Request Lenses

describeStudio_studioId :: Lens' DescribeStudio Text

The Amazon EMR Studio ID.

Destructuring the Response

data DescribeStudioResponse

See: newDescribeStudioResponse smart constructor.

Constructors

DescribeStudioResponse'

Fields

studio :: Maybe Studio

The Amazon EMR Studio details.

httpStatus :: Int

The response's http status code.

newDescribeStudioResponse

Arguments

:: Int
$sel:httpStatus:DescribeStudioResponse'

-> DescribeStudioResponse

Create a value of DescribeStudioResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:studio:DescribeStudioResponse', describeStudioResponse_studio - The Amazon EMR Studio details.

$sel:httpStatus:DescribeStudioResponse', describeStudioResponse_httpStatus - The response's http status code.

Response Lenses

describeStudioResponse_studio :: Lens' DescribeStudioResponse (Maybe Studio)

The Amazon EMR Studio details.

describeStudioResponse_httpStatus :: Lens' DescribeStudioResponse Int

The response's http status code.