Network.AWS.EMR.GetBlockPublicAccessConfiguration

Description

Returns the Amazon EMR block public access configuration for your Amazon Web Services account in the current Region. For more information see Configure Block Public Access for Amazon EMR in the Amazon EMR Management Guide.

Creating a Request

data GetBlockPublicAccessConfiguration

See: newGetBlockPublicAccessConfiguration smart constructor.

Constructors

GetBlockPublicAccessConfiguration'

newGetBlockPublicAccessConfiguration :: GetBlockPublicAccessConfiguration

Create a value of GetBlockPublicAccessConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

Destructuring the Response

data GetBlockPublicAccessConfigurationResponse

See: newGetBlockPublicAccessConfigurationResponse smart constructor.

Constructors

GetBlockPublicAccessConfigurationResponse'

Fields

httpStatus :: Int

The response's http status code.

blockPublicAccessConfiguration :: BlockPublicAccessConfiguration

A configuration for Amazon EMR block public access. The configuration applies to all clusters created in your account for the current Region. The configuration specifies whether block public access is enabled. If block public access is enabled, security groups associated with the cluster cannot have rules that allow inbound traffic from 0.0.0.0/0 or ::/0 on a port, unless the port is specified as an exception using PermittedPublicSecurityGroupRuleRanges in the BlockPublicAccessConfiguration. By default, Port 22 (SSH) is an exception, and public access is allowed on this port. You can change this by updating the block public access configuration to remove the exception.

For accounts that created clusters in a Region before November 25, 2019, block public access is disabled by default in that Region. To use this feature, you must manually enable and configure it. For accounts that did not create an EMR cluster in a Region before this date, block public access is enabled by default in that Region.

blockPublicAccessConfigurationMetadata :: BlockPublicAccessConfigurationMetadata

Properties that describe the Amazon Web Services principal that created the BlockPublicAccessConfiguration using the PutBlockPublicAccessConfiguration action as well as the date and time that the configuration was created. Each time a configuration for block public access is updated, Amazon EMR updates this metadata.

newGetBlockPublicAccessConfigurationResponse

Arguments

:: Int
$sel:httpStatus:GetBlockPublicAccessConfigurationResponse'

-> BlockPublicAccessConfiguration
$sel:blockPublicAccessConfiguration:GetBlockPublicAccessConfigurationResponse'

-> BlockPublicAccessConfigurationMetadata
$sel:blockPublicAccessConfigurationMetadata:GetBlockPublicAccessConfigurationResponse'

-> GetBlockPublicAccessConfigurationResponse

Create a value of GetBlockPublicAccessConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetBlockPublicAccessConfigurationResponse', getBlockPublicAccessConfigurationResponse_httpStatus - The response's http status code.

$sel:blockPublicAccessConfiguration:GetBlockPublicAccessConfigurationResponse', getBlockPublicAccessConfigurationResponse_blockPublicAccessConfiguration - A configuration for Amazon EMR block public access. The configuration applies to all clusters created in your account for the current Region. The configuration specifies whether block public access is enabled. If block public access is enabled, security groups associated with the cluster cannot have rules that allow inbound traffic from 0.0.0.0/0 or ::/0 on a port, unless the port is specified as an exception using PermittedPublicSecurityGroupRuleRanges in the BlockPublicAccessConfiguration. By default, Port 22 (SSH) is an exception, and public access is allowed on this port. You can change this by updating the block public access configuration to remove the exception.

For accounts that created clusters in a Region before November 25, 2019, block public access is disabled by default in that Region. To use this feature, you must manually enable and configure it. For accounts that did not create an EMR cluster in a Region before this date, block public access is enabled by default in that Region.

$sel:blockPublicAccessConfigurationMetadata:GetBlockPublicAccessConfigurationResponse', getBlockPublicAccessConfigurationResponse_blockPublicAccessConfigurationMetadata - Properties that describe the Amazon Web Services principal that created the BlockPublicAccessConfiguration using the PutBlockPublicAccessConfiguration action as well as the date and time that the configuration was created. Each time a configuration for block public access is updated, Amazon EMR updates this metadata.

Response Lenses

getBlockPublicAccessConfigurationResponse_httpStatus :: Lens' GetBlockPublicAccessConfigurationResponse Int

The response's http status code.

getBlockPublicAccessConfigurationResponse_blockPublicAccessConfiguration :: Lens' GetBlockPublicAccessConfigurationResponse BlockPublicAccessConfiguration

A configuration for Amazon EMR block public access. The configuration applies to all clusters created in your account for the current Region. The configuration specifies whether block public access is enabled. If block public access is enabled, security groups associated with the cluster cannot have rules that allow inbound traffic from 0.0.0.0/0 or ::/0 on a port, unless the port is specified as an exception using PermittedPublicSecurityGroupRuleRanges in the BlockPublicAccessConfiguration. By default, Port 22 (SSH) is an exception, and public access is allowed on this port. You can change this by updating the block public access configuration to remove the exception.

For accounts that created clusters in a Region before November 25, 2019, block public access is disabled by default in that Region. To use this feature, you must manually enable and configure it. For accounts that did not create an EMR cluster in a Region before this date, block public access is enabled by default in that Region.

getBlockPublicAccessConfigurationResponse_blockPublicAccessConfigurationMetadata :: Lens' GetBlockPublicAccessConfigurationResponse BlockPublicAccessConfigurationMetadata

Properties that describe the Amazon Web Services principal that created the BlockPublicAccessConfiguration using the PutBlockPublicAccessConfiguration action as well as the date and time that the configuration was created. Each time a configuration for block public access is updated, Amazon EMR updates this metadata.