Network.AWS.EMR.ListInstanceFleets

Description

Lists all available details about the instance fleets in a cluster.

The instance fleet configuration is available only in Amazon EMR versions 4.8.0 and later, excluding 5.0.x versions.

This operation returns paginated results.

Creating a Request

data ListInstanceFleets

See: newListInstanceFleets smart constructor.

Constructors

ListInstanceFleets'

Fields

marker :: Maybe Text

The pagination token that indicates the next set of results to retrieve.

clusterId :: Text

The unique identifier of the cluster.

newListInstanceFleets

Arguments

:: Text
$sel:clusterId:ListInstanceFleets'

-> ListInstanceFleets

Create a value of ListInstanceFleets with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:marker:ListInstanceFleets', listInstanceFleets_marker - The pagination token that indicates the next set of results to retrieve.

$sel:clusterId:ListInstanceFleets', listInstanceFleets_clusterId - The unique identifier of the cluster.

Request Lenses

listInstanceFleets_marker :: Lens' ListInstanceFleets (Maybe Text)

The pagination token that indicates the next set of results to retrieve.

listInstanceFleets_clusterId :: Lens' ListInstanceFleets Text

The unique identifier of the cluster.

Destructuring the Response

data ListInstanceFleetsResponse

See: newListInstanceFleetsResponse smart constructor.

Constructors

ListInstanceFleetsResponse'

Fields

instanceFleets :: Maybe [InstanceFleet]

The list of instance fleets for the cluster and given filters.

marker :: Maybe Text

The pagination token that indicates the next set of results to retrieve.

httpStatus :: Int

The response's http status code.

newListInstanceFleetsResponse

Arguments

:: Int
$sel:httpStatus:ListInstanceFleetsResponse'

-> ListInstanceFleetsResponse

Create a value of ListInstanceFleetsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceFleets:ListInstanceFleetsResponse', listInstanceFleetsResponse_instanceFleets - The list of instance fleets for the cluster and given filters.

$sel:marker:ListInstanceFleets', listInstanceFleetsResponse_marker - The pagination token that indicates the next set of results to retrieve.

$sel:httpStatus:ListInstanceFleetsResponse', listInstanceFleetsResponse_httpStatus - The response's http status code.

Response Lenses

listInstanceFleetsResponse_instanceFleets :: Lens' ListInstanceFleetsResponse (Maybe [InstanceFleet])

The list of instance fleets for the cluster and given filters.

listInstanceFleetsResponse_marker :: Lens' ListInstanceFleetsResponse (Maybe Text)

The pagination token that indicates the next set of results to retrieve.

listInstanceFleetsResponse_httpStatus :: Lens' ListInstanceFleetsResponse Int

The response's http status code.