Network.AWS.EMR.ListNotebookExecutions

Description

Provides summaries of all notebook executions. You can filter the list based on multiple criteria such as status, time range, and editor id. Returns a maximum of 50 notebook executions and a marker to track the paging of a longer notebook execution list across multiple ListNotebookExecution calls.

This operation returns paginated results.

Creating a Request

data ListNotebookExecutions

See: newListNotebookExecutions smart constructor.

Constructors

ListNotebookExecutions'

Fields

status :: Maybe NotebookExecutionStatus

The status filter for listing notebook executions.

START_PENDING indicates that the cluster has received the execution request but execution has not begun.

STARTING indicates that the execution is starting on the cluster.

RUNNING indicates that the execution is being processed by the cluster.

FINISHING indicates that execution processing is in the final stages.

FINISHED indicates that the execution has completed without error.

FAILING indicates that the execution is failing and will not finish successfully.

FAILED indicates that the execution failed.

STOP_PENDING indicates that the cluster has received a StopNotebookExecution request and the stop is pending.

STOPPING indicates that the cluster is in the process of stopping the execution as a result of a StopNotebookExecution request.

STOPPED indicates that the execution stopped because of a StopNotebookExecution request.

editorId :: Maybe Text

The unique ID of the editor associated with the notebook execution.

to :: Maybe POSIX

The end of time range filter for listing notebook executions. The default is the current timestamp.

from :: Maybe POSIX

The beginning of time range filter for listing notebook executions. The default is the timestamp of 30 days ago.

marker :: Maybe Text

The pagination token, returned by a previous ListNotebookExecutions call, that indicates the start of the list for this ListNotebookExecutions call.

newListNotebookExecutions :: ListNotebookExecutions

Create a value of ListNotebookExecutions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ListNotebookExecutions', listNotebookExecutions_status - The status filter for listing notebook executions.

START_PENDING indicates that the cluster has received the execution request but execution has not begun.

STARTING indicates that the execution is starting on the cluster.

RUNNING indicates that the execution is being processed by the cluster.

FINISHING indicates that execution processing is in the final stages.

FINISHED indicates that the execution has completed without error.

FAILING indicates that the execution is failing and will not finish successfully.

FAILED indicates that the execution failed.

STOP_PENDING indicates that the cluster has received a StopNotebookExecution request and the stop is pending.

STOPPING indicates that the cluster is in the process of stopping the execution as a result of a StopNotebookExecution request.

STOPPED indicates that the execution stopped because of a StopNotebookExecution request.

$sel:editorId:ListNotebookExecutions', listNotebookExecutions_editorId - The unique ID of the editor associated with the notebook execution.

$sel:to:ListNotebookExecutions', listNotebookExecutions_to - The end of time range filter for listing notebook executions. The default is the current timestamp.

$sel:from:ListNotebookExecutions', listNotebookExecutions_from - The beginning of time range filter for listing notebook executions. The default is the timestamp of 30 days ago.

$sel:marker:ListNotebookExecutions', listNotebookExecutions_marker - The pagination token, returned by a previous ListNotebookExecutions call, that indicates the start of the list for this ListNotebookExecutions call.

Request Lenses

listNotebookExecutions_status :: Lens' ListNotebookExecutions (Maybe NotebookExecutionStatus)

The status filter for listing notebook executions.

START_PENDING indicates that the cluster has received the execution request but execution has not begun.

STARTING indicates that the execution is starting on the cluster.

RUNNING indicates that the execution is being processed by the cluster.

FINISHING indicates that execution processing is in the final stages.

FINISHED indicates that the execution has completed without error.

FAILING indicates that the execution is failing and will not finish successfully.

FAILED indicates that the execution failed.

STOP_PENDING indicates that the cluster has received a StopNotebookExecution request and the stop is pending.

STOPPING indicates that the cluster is in the process of stopping the execution as a result of a StopNotebookExecution request.

STOPPED indicates that the execution stopped because of a StopNotebookExecution request.

listNotebookExecutions_editorId :: Lens' ListNotebookExecutions (Maybe Text)

The unique ID of the editor associated with the notebook execution.

listNotebookExecutions_to :: Lens' ListNotebookExecutions (Maybe UTCTime)

The end of time range filter for listing notebook executions. The default is the current timestamp.

listNotebookExecutions_from :: Lens' ListNotebookExecutions (Maybe UTCTime)

The beginning of time range filter for listing notebook executions. The default is the timestamp of 30 days ago.

listNotebookExecutions_marker :: Lens' ListNotebookExecutions (Maybe Text)

The pagination token, returned by a previous ListNotebookExecutions call, that indicates the start of the list for this ListNotebookExecutions call.

Destructuring the Response

data ListNotebookExecutionsResponse

See: newListNotebookExecutionsResponse smart constructor.

Constructors

ListNotebookExecutionsResponse'

Fields

notebookExecutions :: Maybe [NotebookExecutionSummary]

A list of notebook executions.

marker :: Maybe Text

A pagination token that a subsequent ListNotebookExecutions can use to determine the next set of results to retrieve.

httpStatus :: Int

The response's http status code.

newListNotebookExecutionsResponse

Arguments

:: Int
$sel:httpStatus:ListNotebookExecutionsResponse'

-> ListNotebookExecutionsResponse

Create a value of ListNotebookExecutionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:notebookExecutions:ListNotebookExecutionsResponse', listNotebookExecutionsResponse_notebookExecutions - A list of notebook executions.

$sel:marker:ListNotebookExecutions', listNotebookExecutionsResponse_marker - A pagination token that a subsequent ListNotebookExecutions can use to determine the next set of results to retrieve.

$sel:httpStatus:ListNotebookExecutionsResponse', listNotebookExecutionsResponse_httpStatus - The response's http status code.

Response Lenses

listNotebookExecutionsResponse_notebookExecutions :: Lens' ListNotebookExecutionsResponse (Maybe [NotebookExecutionSummary])

A list of notebook executions.

listNotebookExecutionsResponse_marker :: Lens' ListNotebookExecutionsResponse (Maybe Text)

A pagination token that a subsequent ListNotebookExecutions can use to determine the next set of results to retrieve.

listNotebookExecutionsResponse_httpStatus :: Lens' ListNotebookExecutionsResponse Int

The response's http status code.