Network.AWS.EMR.ListReleaseLabels

Description

Retrieves release labels of EMR services in the region where the API is called.

Creating a Request

data ListReleaseLabels

See: newListReleaseLabels smart constructor.

Constructors

ListReleaseLabels'

Fields

filters :: Maybe ReleaseLabelFilter

Filters the results of the request. Prefix specifies the prefix of release labels to return. Application specifies the application (with/without version) of release labels to return.

nextToken :: Maybe Text

Specifies the next page of results. If NextToken is not specified, which is usually the case for the first request of ListReleaseLabels, the first page of results are determined by other filtering parameters or by the latest version. The ListReleaseLabels request fails if the identity (Amazon Web Services account ID) and all filtering parameters are different from the original request, or if the NextToken is expired or tampered with.

maxResults :: Maybe Natural

Defines the maximum number of release labels to return in a single response. The default is 100.

newListReleaseLabels :: ListReleaseLabels

Create a value of ListReleaseLabels with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:ListReleaseLabels', listReleaseLabels_filters - Filters the results of the request. Prefix specifies the prefix of release labels to return. Application specifies the application (with/without version) of release labels to return.

$sel:nextToken:ListReleaseLabels', listReleaseLabels_nextToken - Specifies the next page of results. If NextToken is not specified, which is usually the case for the first request of ListReleaseLabels, the first page of results are determined by other filtering parameters or by the latest version. The ListReleaseLabels request fails if the identity (Amazon Web Services account ID) and all filtering parameters are different from the original request, or if the NextToken is expired or tampered with.

$sel:maxResults:ListReleaseLabels', listReleaseLabels_maxResults - Defines the maximum number of release labels to return in a single response. The default is 100.

Request Lenses

listReleaseLabels_filters :: Lens' ListReleaseLabels (Maybe ReleaseLabelFilter)

Filters the results of the request. Prefix specifies the prefix of release labels to return. Application specifies the application (with/without version) of release labels to return.

listReleaseLabels_nextToken :: Lens' ListReleaseLabels (Maybe Text)

Specifies the next page of results. If NextToken is not specified, which is usually the case for the first request of ListReleaseLabels, the first page of results are determined by other filtering parameters or by the latest version. The ListReleaseLabels request fails if the identity (Amazon Web Services account ID) and all filtering parameters are different from the original request, or if the NextToken is expired or tampered with.

listReleaseLabels_maxResults :: Lens' ListReleaseLabels (Maybe Natural)

Defines the maximum number of release labels to return in a single response. The default is 100.

Destructuring the Response

data ListReleaseLabelsResponse

See: newListReleaseLabelsResponse smart constructor.

Constructors

ListReleaseLabelsResponse'

Fields

releaseLabels :: Maybe [Text]

The returned release labels.

nextToken :: Maybe Text

Used to paginate the next page of results if specified in the next ListReleaseLabels request.

httpStatus :: Int

The response's http status code.

newListReleaseLabelsResponse

Arguments

:: Int
$sel:httpStatus:ListReleaseLabelsResponse'

-> ListReleaseLabelsResponse

Create a value of ListReleaseLabelsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:releaseLabels:ListReleaseLabelsResponse', listReleaseLabelsResponse_releaseLabels - The returned release labels.

$sel:nextToken:ListReleaseLabels', listReleaseLabelsResponse_nextToken - Used to paginate the next page of results if specified in the next ListReleaseLabels request.

$sel:httpStatus:ListReleaseLabelsResponse', listReleaseLabelsResponse_httpStatus - The response's http status code.

Response Lenses

listReleaseLabelsResponse_releaseLabels :: Lens' ListReleaseLabelsResponse (Maybe [Text])

The returned release labels.

listReleaseLabelsResponse_nextToken :: Lens' ListReleaseLabelsResponse (Maybe Text)

Used to paginate the next page of results if specified in the next ListReleaseLabels request.

listReleaseLabelsResponse_httpStatus :: Lens' ListReleaseLabelsResponse Int

The response's http status code.