Network.AWS.EMR.ListStudios

Description

Returns a list of all Amazon EMR Studios associated with the Amazon Web Services account. The list includes details such as ID, Studio Access URL, and creation time for each Studio.

This operation returns paginated results.

Creating a Request

data ListStudios

See: newListStudios smart constructor.

Constructors

ListStudios'

Fields

marker :: Maybe Text

The pagination token that indicates the set of results to retrieve.

newListStudios :: ListStudios

Create a value of ListStudios with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:marker:ListStudios', listStudios_marker - The pagination token that indicates the set of results to retrieve.

Request Lenses

listStudios_marker :: Lens' ListStudios (Maybe Text)

The pagination token that indicates the set of results to retrieve.

Destructuring the Response

data ListStudiosResponse

See: newListStudiosResponse smart constructor.

Constructors

ListStudiosResponse'

Fields

studios :: Maybe [StudioSummary]

The list of Studio summary objects.

marker :: Maybe Text

The pagination token that indicates the next set of results to retrieve.

httpStatus :: Int

The response's http status code.

newListStudiosResponse

Arguments

:: Int
$sel:httpStatus:ListStudiosResponse'

-> ListStudiosResponse

Create a value of ListStudiosResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:studios:ListStudiosResponse', listStudiosResponse_studios - The list of Studio summary objects.

$sel:marker:ListStudios', listStudiosResponse_marker - The pagination token that indicates the next set of results to retrieve.

$sel:httpStatus:ListStudiosResponse', listStudiosResponse_httpStatus - The response's http status code.

Response Lenses

listStudiosResponse_studios :: Lens' ListStudiosResponse (Maybe [StudioSummary])

The list of Studio summary objects.

listStudiosResponse_marker :: Lens' ListStudiosResponse (Maybe Text)

The pagination token that indicates the next set of results to retrieve.

listStudiosResponse_httpStatus :: Lens' ListStudiosResponse Int

The response's http status code.