Network.AWS.EMR.PutAutoTerminationPolicy

Description

Creates or updates an auto-termination policy for an Amazon EMR cluster. An auto-termination policy defines the amount of idle time in seconds after which a cluster automatically terminates. For alternative cluster termination options, see Control cluster termination.

Creating a Request

data PutAutoTerminationPolicy

See: newPutAutoTerminationPolicy smart constructor.

Constructors

PutAutoTerminationPolicy'

Fields

autoTerminationPolicy :: Maybe AutoTerminationPolicy

Specifies the auto-termination policy to attach to the cluster.

clusterId :: Text

Specifies the ID of the Amazon EMR cluster to which the auto-termination policy will be attached.

newPutAutoTerminationPolicy

Arguments

:: Text
$sel:clusterId:PutAutoTerminationPolicy'

-> PutAutoTerminationPolicy

Create a value of PutAutoTerminationPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:autoTerminationPolicy:PutAutoTerminationPolicy', putAutoTerminationPolicy_autoTerminationPolicy - Specifies the auto-termination policy to attach to the cluster.

$sel:clusterId:PutAutoTerminationPolicy', putAutoTerminationPolicy_clusterId - Specifies the ID of the Amazon EMR cluster to which the auto-termination policy will be attached.

Request Lenses

putAutoTerminationPolicy_autoTerminationPolicy :: Lens' PutAutoTerminationPolicy (Maybe AutoTerminationPolicy)

Specifies the auto-termination policy to attach to the cluster.

putAutoTerminationPolicy_clusterId :: Lens' PutAutoTerminationPolicy Text

Specifies the ID of the Amazon EMR cluster to which the auto-termination policy will be attached.

Destructuring the Response

data PutAutoTerminationPolicyResponse

See: newPutAutoTerminationPolicyResponse smart constructor.

Constructors

PutAutoTerminationPolicyResponse'

Fields

httpStatus :: Int

The response's http status code.

newPutAutoTerminationPolicyResponse

Arguments

:: Int
$sel:httpStatus:PutAutoTerminationPolicyResponse'

-> PutAutoTerminationPolicyResponse

Create a value of PutAutoTerminationPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:PutAutoTerminationPolicyResponse', putAutoTerminationPolicyResponse_httpStatus - The response's http status code.

Response Lenses

putAutoTerminationPolicyResponse_httpStatus :: Lens' PutAutoTerminationPolicyResponse Int

The response's http status code.