Network.AWS.EMR.PutBlockPublicAccessConfiguration

Description

Creates or updates an Amazon EMR block public access configuration for your Amazon Web Services account in the current Region. For more information see Configure Block Public Access for Amazon EMR in the Amazon EMR Management Guide.

Creating a Request

data PutBlockPublicAccessConfiguration

See: newPutBlockPublicAccessConfiguration smart constructor.

Constructors

PutBlockPublicAccessConfiguration'

Fields

blockPublicAccessConfiguration :: BlockPublicAccessConfiguration

A configuration for Amazon EMR block public access. The configuration applies to all clusters created in your account for the current Region. The configuration specifies whether block public access is enabled. If block public access is enabled, security groups associated with the cluster cannot have rules that allow inbound traffic from 0.0.0.0/0 or ::/0 on a port, unless the port is specified as an exception using PermittedPublicSecurityGroupRuleRanges in the BlockPublicAccessConfiguration. By default, Port 22 (SSH) is an exception, and public access is allowed on this port. You can change this by updating BlockPublicSecurityGroupRules to remove the exception.

For accounts that created clusters in a Region before November 25, 2019, block public access is disabled by default in that Region. To use this feature, you must manually enable and configure it. For accounts that did not create an EMR cluster in a Region before this date, block public access is enabled by default in that Region.

newPutBlockPublicAccessConfiguration

Arguments

:: BlockPublicAccessConfiguration
$sel:blockPublicAccessConfiguration:PutBlockPublicAccessConfiguration'

-> PutBlockPublicAccessConfiguration

Create a value of PutBlockPublicAccessConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:blockPublicAccessConfiguration:PutBlockPublicAccessConfiguration', putBlockPublicAccessConfiguration_blockPublicAccessConfiguration - A configuration for Amazon EMR block public access. The configuration applies to all clusters created in your account for the current Region. The configuration specifies whether block public access is enabled. If block public access is enabled, security groups associated with the cluster cannot have rules that allow inbound traffic from 0.0.0.0/0 or ::/0 on a port, unless the port is specified as an exception using PermittedPublicSecurityGroupRuleRanges in the BlockPublicAccessConfiguration. By default, Port 22 (SSH) is an exception, and public access is allowed on this port. You can change this by updating BlockPublicSecurityGroupRules to remove the exception.

For accounts that created clusters in a Region before November 25, 2019, block public access is disabled by default in that Region. To use this feature, you must manually enable and configure it. For accounts that did not create an EMR cluster in a Region before this date, block public access is enabled by default in that Region.

Request Lenses

putBlockPublicAccessConfiguration_blockPublicAccessConfiguration :: Lens' PutBlockPublicAccessConfiguration BlockPublicAccessConfiguration

A configuration for Amazon EMR block public access. The configuration applies to all clusters created in your account for the current Region. The configuration specifies whether block public access is enabled. If block public access is enabled, security groups associated with the cluster cannot have rules that allow inbound traffic from 0.0.0.0/0 or ::/0 on a port, unless the port is specified as an exception using PermittedPublicSecurityGroupRuleRanges in the BlockPublicAccessConfiguration. By default, Port 22 (SSH) is an exception, and public access is allowed on this port. You can change this by updating BlockPublicSecurityGroupRules to remove the exception.

For accounts that created clusters in a Region before November 25, 2019, block public access is disabled by default in that Region. To use this feature, you must manually enable and configure it. For accounts that did not create an EMR cluster in a Region before this date, block public access is enabled by default in that Region.

Destructuring the Response

data PutBlockPublicAccessConfigurationResponse

See: newPutBlockPublicAccessConfigurationResponse smart constructor.

Constructors

PutBlockPublicAccessConfigurationResponse'

Fields

httpStatus :: Int

The response's http status code.

newPutBlockPublicAccessConfigurationResponse

Arguments

:: Int
$sel:httpStatus:PutBlockPublicAccessConfigurationResponse'

-> PutBlockPublicAccessConfigurationResponse

Create a value of PutBlockPublicAccessConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:PutBlockPublicAccessConfigurationResponse', putBlockPublicAccessConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

putBlockPublicAccessConfigurationResponse_httpStatus :: Lens' PutBlockPublicAccessConfigurationResponse Int

The response's http status code.