Network.AWS.EMR.RunJobFlow

Description

RunJobFlow creates and starts running a new cluster (job flow). The cluster runs the steps specified. After the steps complete, the cluster stops and the HDFS partition is lost. To prevent loss of data, configure the last step of the job flow to store results in Amazon S3. If the JobFlowInstancesConfig KeepJobFlowAliveWhenNoSteps parameter is set to TRUE, the cluster transitions to the WAITING state rather than shutting down after the steps have completed.

For additional protection, you can set the JobFlowInstancesConfig TerminationProtected parameter to TRUE to lock the cluster and prevent it from being terminated by API call, user intervention, or in the event of a job flow error.

A maximum of 256 steps are allowed in each job flow.

If your cluster is long-running (such as a Hive data warehouse) or complex, you may require more than 256 steps to process your data. You can bypass the 256-step limitation in various ways, including using the SSH shell to connect to the master node and submitting queries directly to the software running on the master node, such as Hive and Hadoop. For more information on how to do this, see Add More than 256 Steps to a Cluster in the Amazon EMR Management Guide.

For long running clusters, we recommend that you periodically store your results.

The instance fleets configuration is available only in Amazon EMR versions 4.8.0 and later, excluding 5.0.x versions. The RunJobFlow request can contain InstanceFleets parameters or InstanceGroups parameters, but not both.

Creating a Request

data RunJobFlow

Input to the RunJobFlow operation.

See: newRunJobFlow smart constructor.

Constructors

RunJobFlow'

Fields

logEncryptionKmsKeyId :: Maybe Text

The KMS key used for encrypting log files. If a value is not provided, the logs remain encrypted by AES-256. This attribute is only available with Amazon EMR version 5.30.0 and later, excluding Amazon EMR 6.0.0.

amiVersion :: Maybe Text

Applies only to Amazon EMR AMI versions 3.x and 2.x. For Amazon EMR releases 4.0 and later, ReleaseLabel is used. To specify a custom AMI, use CustomAmiID.

ebsRootVolumeSize :: Maybe Int

The size, in GiB, of the Amazon EBS root device volume of the Linux AMI that is used for each EC2 instance. Available in Amazon EMR version 4.x and later.

additionalInfo :: Maybe Text

A JSON string for selecting additional features.

autoTerminationPolicy :: Maybe AutoTerminationPolicy

configurations :: Maybe [Configuration]

For Amazon EMR releases 4.0 and later. The list of configurations supplied for the EMR cluster you are creating.

customAmiId :: Maybe Text

Available only in Amazon EMR version 5.7.0 and later. The ID of a custom Amazon EBS-backed Linux AMI. If specified, Amazon EMR uses this AMI when it launches cluster EC2 instances. For more information about custom AMIs in Amazon EMR, see Using a Custom AMI in the Amazon EMR Management Guide. If omitted, the cluster uses the base Linux AMI for the ReleaseLabel specified. For Amazon EMR versions 2.x and 3.x, use AmiVersion instead.

For information about creating a custom AMI, see Creating an Amazon EBS-Backed Linux AMI in the Amazon Elastic Compute Cloud User Guide for Linux Instances. For information about finding an AMI ID, see Finding a Linux AMI.

autoScalingRole :: Maybe Text

An IAM role for automatic scaling policies. The default role is EMR_AutoScaling_DefaultRole. The IAM role provides permissions that the automatic scaling feature requires to launch and terminate EC2 instances in an instance group.

securityConfiguration :: Maybe Text

The name of a security configuration to apply to the cluster.

scaleDownBehavior :: Maybe ScaleDownBehavior

Specifies the way that individual Amazon EC2 instances terminate when an automatic scale-in activity occurs or an instance group is resized. TERMINATE_AT_INSTANCE_HOUR indicates that Amazon EMR terminates nodes at the instance-hour boundary, regardless of when the request to terminate the instance was submitted. This option is only available with Amazon EMR 5.1.0 and later and is the default for clusters created using that version. TERMINATE_AT_TASK_COMPLETION indicates that Amazon EMR adds nodes to a deny list and drains tasks from nodes before terminating the Amazon EC2 instances, regardless of the instance-hour boundary. With either behavior, Amazon EMR removes the least active nodes first and blocks instance termination if it could lead to HDFS corruption. TERMINATE_AT_TASK_COMPLETION available only in Amazon EMR version 4.1.0 and later, and is the default for versions of Amazon EMR earlier than 5.1.0.

steps :: Maybe [StepConfig]

A list of steps to run.

jobFlowRole :: Maybe Text

Also called instance profile and EC2 role. An IAM role for an EMR cluster. The EC2 instances of the cluster assume this role. The default role is EMR_EC2_DefaultRole. In order to use the default role, you must have already created it using the CLI or console.

bootstrapActions :: Maybe [BootstrapActionConfig]

A list of bootstrap actions to run before Hadoop starts on the cluster nodes.

releaseLabel :: Maybe Text

The Amazon EMR release label, which determines the version of open-source application packages installed on the cluster. Release labels are in the form emr-x.x.x, where x.x.x is an Amazon EMR release version such as emr-5.14.0. For more information about Amazon EMR release versions and included application versions and features, see https://docs.aws.amazon.com/emr/latest/ReleaseGuide/. The release label applies only to Amazon EMR releases version 4.0 and later. Earlier versions use AmiVersion.

repoUpgradeOnBoot :: Maybe RepoUpgradeOnBoot

Applies only when CustomAmiID is used. Specifies which updates from the Amazon Linux AMI package repositories to apply automatically when the instance boots using the AMI. If omitted, the default is SECURITY, which indicates that only security updates are applied. If NONE is specified, no updates are applied, and all updates must be applied manually.

placementGroupConfigs :: Maybe [PlacementGroupConfig]

The specified placement group configuration for an Amazon EMR cluster.

logUri :: Maybe Text

The location in Amazon S3 to write the log files of the job flow. If a value is not provided, logs are not created.

kerberosAttributes :: Maybe KerberosAttributes

Attributes for Kerberos configuration when Kerberos authentication is enabled using a security configuration. For more information see Use Kerberos Authentication in the Amazon EMR Management Guide.

newSupportedProducts' :: Maybe [SupportedProductConfig]

For Amazon EMR releases 3.x and 2.x. For Amazon EMR releases 4.x and later, use Applications.

A list of strings that indicates third-party software to use with the job flow that accepts a user argument list. EMR accepts and forwards the argument list to the corresponding installation script as bootstrap action arguments. For more information, see "Launch a Job Flow on the MapR Distribution for Hadoop" in the Amazon EMR Developer Guide. Supported values are:

"mapr-m3" - launch the cluster using MapR M3 Edition.

"mapr-m5" - launch the cluster using MapR M5 Edition.

"mapr" with the user arguments specifying "--edition,m3" or "--edition,m5" - launch the job flow using MapR M3 or M5 Edition respectively.

"mapr-m7" - launch the cluster using MapR M7 Edition.

"hunk" - launch the cluster with the Hunk Big Data Analytics Platform.

"hue"- launch the cluster with Hue installed.

"spark" - launch the cluster with Apache Spark installed.

"ganglia" - launch the cluster with the Ganglia Monitoring System installed.

managedScalingPolicy :: Maybe ManagedScalingPolicy

The specified managed scaling policy for an Amazon EMR cluster.

visibleToAllUsers :: Maybe Bool

Set this value to true so that IAM principals in the Amazon Web Services account associated with the cluster can perform EMR actions on the cluster that their IAM policies allow. This value defaults to true for clusters created using the EMR API or the CLI create-cluster command.

When set to false, only the IAM principal that created the cluster and the Amazon Web Services account root user can perform EMR actions for the cluster, regardless of the IAM permissions policies attached to other IAM principals. For more information, see Understanding the EMR Cluster VisibleToAllUsers Setting in the Amazon EMRManagement Guide.

supportedProducts :: Maybe [Text]

For Amazon EMR releases 3.x and 2.x. For Amazon EMR releases 4.x and later, use Applications.

A list of strings that indicates third-party software to use. For more information, see the Amazon EMR Developer Guide. Currently supported values are:

"mapr-m3" - launch the job flow using MapR M3 Edition.

"mapr-m5" - launch the job flow using MapR M5 Edition.

stepConcurrencyLevel :: Maybe Int

Specifies the number of steps that can be executed concurrently. The default value is 1. The maximum value is 256.

applications :: Maybe [Application]

Applies to Amazon EMR releases 4.0 and later. A case-insensitive list of applications for Amazon EMR to install and configure when launching the cluster. For a list of applications available for each Amazon EMR release version, see the Amazon EMR Release Guide.

tags :: Maybe [Tag]

A list of tags to associate with a cluster and propagate to Amazon EC2 instances.

serviceRole :: Maybe Text

The IAM role that Amazon EMR assumes in order to access Amazon Web Services resources on your behalf.

name :: Text

The name of the job flow.

instances :: JobFlowInstancesConfig

A specification of the number and type of Amazon EC2 instances.

newRunJobFlow

Arguments

:: Text
$sel:name:RunJobFlow'

-> JobFlowInstancesConfig
$sel:instances:RunJobFlow'

-> RunJobFlow

Create a value of RunJobFlow with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:logEncryptionKmsKeyId:RunJobFlow', runJobFlow_logEncryptionKmsKeyId - The KMS key used for encrypting log files. If a value is not provided, the logs remain encrypted by AES-256. This attribute is only available with Amazon EMR version 5.30.0 and later, excluding Amazon EMR 6.0.0.

$sel:amiVersion:RunJobFlow', runJobFlow_amiVersion - Applies only to Amazon EMR AMI versions 3.x and 2.x. For Amazon EMR releases 4.0 and later, ReleaseLabel is used. To specify a custom AMI, use CustomAmiID.

$sel:ebsRootVolumeSize:RunJobFlow', runJobFlow_ebsRootVolumeSize - The size, in GiB, of the Amazon EBS root device volume of the Linux AMI that is used for each EC2 instance. Available in Amazon EMR version 4.x and later.

$sel:additionalInfo:RunJobFlow', runJobFlow_additionalInfo - A JSON string for selecting additional features.

$sel:autoTerminationPolicy:RunJobFlow', runJobFlow_autoTerminationPolicy - Undocumented member.

$sel:configurations:RunJobFlow', runJobFlow_configurations - For Amazon EMR releases 4.0 and later. The list of configurations supplied for the EMR cluster you are creating.

$sel:customAmiId:RunJobFlow', runJobFlow_customAmiId - Available only in Amazon EMR version 5.7.0 and later. The ID of a custom Amazon EBS-backed Linux AMI. If specified, Amazon EMR uses this AMI when it launches cluster EC2 instances. For more information about custom AMIs in Amazon EMR, see Using a Custom AMI in the Amazon EMR Management Guide. If omitted, the cluster uses the base Linux AMI for the ReleaseLabel specified. For Amazon EMR versions 2.x and 3.x, use AmiVersion instead.

For information about creating a custom AMI, see Creating an Amazon EBS-Backed Linux AMI in the Amazon Elastic Compute Cloud User Guide for Linux Instances. For information about finding an AMI ID, see Finding a Linux AMI.

$sel:autoScalingRole:RunJobFlow', runJobFlow_autoScalingRole - An IAM role for automatic scaling policies. The default role is EMR_AutoScaling_DefaultRole. The IAM role provides permissions that the automatic scaling feature requires to launch and terminate EC2 instances in an instance group.

$sel:securityConfiguration:RunJobFlow', runJobFlow_securityConfiguration - The name of a security configuration to apply to the cluster.

$sel:scaleDownBehavior:RunJobFlow', runJobFlow_scaleDownBehavior - Specifies the way that individual Amazon EC2 instances terminate when an automatic scale-in activity occurs or an instance group is resized. TERMINATE_AT_INSTANCE_HOUR indicates that Amazon EMR terminates nodes at the instance-hour boundary, regardless of when the request to terminate the instance was submitted. This option is only available with Amazon EMR 5.1.0 and later and is the default for clusters created using that version. TERMINATE_AT_TASK_COMPLETION indicates that Amazon EMR adds nodes to a deny list and drains tasks from nodes before terminating the Amazon EC2 instances, regardless of the instance-hour boundary. With either behavior, Amazon EMR removes the least active nodes first and blocks instance termination if it could lead to HDFS corruption. TERMINATE_AT_TASK_COMPLETION available only in Amazon EMR version 4.1.0 and later, and is the default for versions of Amazon EMR earlier than 5.1.0.

$sel:steps:RunJobFlow', runJobFlow_steps - A list of steps to run.

$sel:jobFlowRole:RunJobFlow', runJobFlow_jobFlowRole - Also called instance profile and EC2 role. An IAM role for an EMR cluster. The EC2 instances of the cluster assume this role. The default role is EMR_EC2_DefaultRole. In order to use the default role, you must have already created it using the CLI or console.

$sel:bootstrapActions:RunJobFlow', runJobFlow_bootstrapActions - A list of bootstrap actions to run before Hadoop starts on the cluster nodes.

$sel:releaseLabel:RunJobFlow', runJobFlow_releaseLabel - The Amazon EMR release label, which determines the version of open-source application packages installed on the cluster. Release labels are in the form emr-x.x.x, where x.x.x is an Amazon EMR release version such as emr-5.14.0. For more information about Amazon EMR release versions and included application versions and features, see https://docs.aws.amazon.com/emr/latest/ReleaseGuide/. The release label applies only to Amazon EMR releases version 4.0 and later. Earlier versions use AmiVersion.

$sel:repoUpgradeOnBoot:RunJobFlow', runJobFlow_repoUpgradeOnBoot - Applies only when CustomAmiID is used. Specifies which updates from the Amazon Linux AMI package repositories to apply automatically when the instance boots using the AMI. If omitted, the default is SECURITY, which indicates that only security updates are applied. If NONE is specified, no updates are applied, and all updates must be applied manually.

$sel:placementGroupConfigs:RunJobFlow', runJobFlow_placementGroupConfigs - The specified placement group configuration for an Amazon EMR cluster.

$sel:logUri:RunJobFlow', runJobFlow_logUri - The location in Amazon S3 to write the log files of the job flow. If a value is not provided, logs are not created.

$sel:kerberosAttributes:RunJobFlow', runJobFlow_kerberosAttributes - Attributes for Kerberos configuration when Kerberos authentication is enabled using a security configuration. For more information see Use Kerberos Authentication in the Amazon EMR Management Guide.

$sel:newSupportedProducts':RunJobFlow', runJobFlow_newSupportedProducts - For Amazon EMR releases 3.x and 2.x. For Amazon EMR releases 4.x and later, use Applications.

A list of strings that indicates third-party software to use with the job flow that accepts a user argument list. EMR accepts and forwards the argument list to the corresponding installation script as bootstrap action arguments. For more information, see "Launch a Job Flow on the MapR Distribution for Hadoop" in the Amazon EMR Developer Guide. Supported values are:

"mapr-m3" - launch the cluster using MapR M3 Edition.

"mapr-m5" - launch the cluster using MapR M5 Edition.

"mapr" with the user arguments specifying "--edition,m3" or "--edition,m5" - launch the job flow using MapR M3 or M5 Edition respectively.

"mapr-m7" - launch the cluster using MapR M7 Edition.

"hunk" - launch the cluster with the Hunk Big Data Analytics Platform.

"hue"- launch the cluster with Hue installed.

"spark" - launch the cluster with Apache Spark installed.

"ganglia" - launch the cluster with the Ganglia Monitoring System installed.

$sel:managedScalingPolicy:RunJobFlow', runJobFlow_managedScalingPolicy - The specified managed scaling policy for an Amazon EMR cluster.

$sel:visibleToAllUsers:RunJobFlow', runJobFlow_visibleToAllUsers - Set this value to true so that IAM principals in the Amazon Web Services account associated with the cluster can perform EMR actions on the cluster that their IAM policies allow. This value defaults to true for clusters created using the EMR API or the CLI create-cluster command.

When set to false, only the IAM principal that created the cluster and the Amazon Web Services account root user can perform EMR actions for the cluster, regardless of the IAM permissions policies attached to other IAM principals. For more information, see Understanding the EMR Cluster VisibleToAllUsers Setting in the Amazon EMRManagement Guide.

$sel:supportedProducts:RunJobFlow', runJobFlow_supportedProducts - For Amazon EMR releases 3.x and 2.x. For Amazon EMR releases 4.x and later, use Applications.

A list of strings that indicates third-party software to use. For more information, see the Amazon EMR Developer Guide. Currently supported values are:

"mapr-m3" - launch the job flow using MapR M3 Edition.

"mapr-m5" - launch the job flow using MapR M5 Edition.

$sel:stepConcurrencyLevel:RunJobFlow', runJobFlow_stepConcurrencyLevel - Specifies the number of steps that can be executed concurrently. The default value is 1. The maximum value is 256.

$sel:applications:RunJobFlow', runJobFlow_applications - Applies to Amazon EMR releases 4.0 and later. A case-insensitive list of applications for Amazon EMR to install and configure when launching the cluster. For a list of applications available for each Amazon EMR release version, see the Amazon EMR Release Guide.

$sel:tags:RunJobFlow', runJobFlow_tags - A list of tags to associate with a cluster and propagate to Amazon EC2 instances.

$sel:serviceRole:RunJobFlow', runJobFlow_serviceRole - The IAM role that Amazon EMR assumes in order to access Amazon Web Services resources on your behalf.

$sel:name:RunJobFlow', runJobFlow_name - The name of the job flow.

$sel:instances:RunJobFlow', runJobFlow_instances - A specification of the number and type of Amazon EC2 instances.

Request Lenses

runJobFlow_logEncryptionKmsKeyId :: Lens' RunJobFlow (Maybe Text)

The KMS key used for encrypting log files. If a value is not provided, the logs remain encrypted by AES-256. This attribute is only available with Amazon EMR version 5.30.0 and later, excluding Amazon EMR 6.0.0.

runJobFlow_amiVersion :: Lens' RunJobFlow (Maybe Text)

Applies only to Amazon EMR AMI versions 3.x and 2.x. For Amazon EMR releases 4.0 and later, ReleaseLabel is used. To specify a custom AMI, use CustomAmiID.

runJobFlow_ebsRootVolumeSize :: Lens' RunJobFlow (Maybe Int)

The size, in GiB, of the Amazon EBS root device volume of the Linux AMI that is used for each EC2 instance. Available in Amazon EMR version 4.x and later.

runJobFlow_additionalInfo :: Lens' RunJobFlow (Maybe Text)

A JSON string for selecting additional features.

runJobFlow_autoTerminationPolicy :: Lens' RunJobFlow (Maybe AutoTerminationPolicy)

Undocumented member.

runJobFlow_configurations :: Lens' RunJobFlow (Maybe [Configuration])

For Amazon EMR releases 4.0 and later. The list of configurations supplied for the EMR cluster you are creating.

runJobFlow_customAmiId :: Lens' RunJobFlow (Maybe Text)

Available only in Amazon EMR version 5.7.0 and later. The ID of a custom Amazon EBS-backed Linux AMI. If specified, Amazon EMR uses this AMI when it launches cluster EC2 instances. For more information about custom AMIs in Amazon EMR, see Using a Custom AMI in the Amazon EMR Management Guide. If omitted, the cluster uses the base Linux AMI for the ReleaseLabel specified. For Amazon EMR versions 2.x and 3.x, use AmiVersion instead.

For information about creating a custom AMI, see Creating an Amazon EBS-Backed Linux AMI in the Amazon Elastic Compute Cloud User Guide for Linux Instances. For information about finding an AMI ID, see Finding a Linux AMI.

runJobFlow_autoScalingRole :: Lens' RunJobFlow (Maybe Text)

An IAM role for automatic scaling policies. The default role is EMR_AutoScaling_DefaultRole. The IAM role provides permissions that the automatic scaling feature requires to launch and terminate EC2 instances in an instance group.

runJobFlow_securityConfiguration :: Lens' RunJobFlow (Maybe Text)

The name of a security configuration to apply to the cluster.

runJobFlow_scaleDownBehavior :: Lens' RunJobFlow (Maybe ScaleDownBehavior)

Specifies the way that individual Amazon EC2 instances terminate when an automatic scale-in activity occurs or an instance group is resized. TERMINATE_AT_INSTANCE_HOUR indicates that Amazon EMR terminates nodes at the instance-hour boundary, regardless of when the request to terminate the instance was submitted. This option is only available with Amazon EMR 5.1.0 and later and is the default for clusters created using that version. TERMINATE_AT_TASK_COMPLETION indicates that Amazon EMR adds nodes to a deny list and drains tasks from nodes before terminating the Amazon EC2 instances, regardless of the instance-hour boundary. With either behavior, Amazon EMR removes the least active nodes first and blocks instance termination if it could lead to HDFS corruption. TERMINATE_AT_TASK_COMPLETION available only in Amazon EMR version 4.1.0 and later, and is the default for versions of Amazon EMR earlier than 5.1.0.

runJobFlow_steps :: Lens' RunJobFlow (Maybe [StepConfig])

A list of steps to run.

runJobFlow_jobFlowRole :: Lens' RunJobFlow (Maybe Text)

Also called instance profile and EC2 role. An IAM role for an EMR cluster. The EC2 instances of the cluster assume this role. The default role is EMR_EC2_DefaultRole. In order to use the default role, you must have already created it using the CLI or console.

runJobFlow_bootstrapActions :: Lens' RunJobFlow (Maybe [BootstrapActionConfig])

A list of bootstrap actions to run before Hadoop starts on the cluster nodes.

runJobFlow_releaseLabel :: Lens' RunJobFlow (Maybe Text)

The Amazon EMR release label, which determines the version of open-source application packages installed on the cluster. Release labels are in the form emr-x.x.x, where x.x.x is an Amazon EMR release version such as emr-5.14.0. For more information about Amazon EMR release versions and included application versions and features, see https://docs.aws.amazon.com/emr/latest/ReleaseGuide/. The release label applies only to Amazon EMR releases version 4.0 and later. Earlier versions use AmiVersion.

runJobFlow_repoUpgradeOnBoot :: Lens' RunJobFlow (Maybe RepoUpgradeOnBoot)

Applies only when CustomAmiID is used. Specifies which updates from the Amazon Linux AMI package repositories to apply automatically when the instance boots using the AMI. If omitted, the default is SECURITY, which indicates that only security updates are applied. If NONE is specified, no updates are applied, and all updates must be applied manually.

runJobFlow_placementGroupConfigs :: Lens' RunJobFlow (Maybe [PlacementGroupConfig])

The specified placement group configuration for an Amazon EMR cluster.

runJobFlow_logUri :: Lens' RunJobFlow (Maybe Text)

The location in Amazon S3 to write the log files of the job flow. If a value is not provided, logs are not created.

runJobFlow_kerberosAttributes :: Lens' RunJobFlow (Maybe KerberosAttributes)

Attributes for Kerberos configuration when Kerberos authentication is enabled using a security configuration. For more information see Use Kerberos Authentication in the Amazon EMR Management Guide.

runJobFlow_newSupportedProducts :: Lens' RunJobFlow (Maybe [SupportedProductConfig])

For Amazon EMR releases 3.x and 2.x. For Amazon EMR releases 4.x and later, use Applications.

A list of strings that indicates third-party software to use with the job flow that accepts a user argument list. EMR accepts and forwards the argument list to the corresponding installation script as bootstrap action arguments. For more information, see "Launch a Job Flow on the MapR Distribution for Hadoop" in the Amazon EMR Developer Guide. Supported values are:

"mapr-m3" - launch the cluster using MapR M3 Edition.

"mapr-m5" - launch the cluster using MapR M5 Edition.

"mapr" with the user arguments specifying "--edition,m3" or "--edition,m5" - launch the job flow using MapR M3 or M5 Edition respectively.

"mapr-m7" - launch the cluster using MapR M7 Edition.

"hunk" - launch the cluster with the Hunk Big Data Analytics Platform.

"hue"- launch the cluster with Hue installed.

"spark" - launch the cluster with Apache Spark installed.

"ganglia" - launch the cluster with the Ganglia Monitoring System installed.

runJobFlow_managedScalingPolicy :: Lens' RunJobFlow (Maybe ManagedScalingPolicy)

The specified managed scaling policy for an Amazon EMR cluster.

runJobFlow_visibleToAllUsers :: Lens' RunJobFlow (Maybe Bool)

Set this value to true so that IAM principals in the Amazon Web Services account associated with the cluster can perform EMR actions on the cluster that their IAM policies allow. This value defaults to true for clusters created using the EMR API or the CLI create-cluster command.

When set to false, only the IAM principal that created the cluster and the Amazon Web Services account root user can perform EMR actions for the cluster, regardless of the IAM permissions policies attached to other IAM principals. For more information, see Understanding the EMR Cluster VisibleToAllUsers Setting in the Amazon EMRManagement Guide.

runJobFlow_supportedProducts :: Lens' RunJobFlow (Maybe [Text])

For Amazon EMR releases 3.x and 2.x. For Amazon EMR releases 4.x and later, use Applications.

A list of strings that indicates third-party software to use. For more information, see the Amazon EMR Developer Guide. Currently supported values are:

"mapr-m3" - launch the job flow using MapR M3 Edition.

"mapr-m5" - launch the job flow using MapR M5 Edition.

runJobFlow_stepConcurrencyLevel :: Lens' RunJobFlow (Maybe Int)

Specifies the number of steps that can be executed concurrently. The default value is 1. The maximum value is 256.

runJobFlow_applications :: Lens' RunJobFlow (Maybe [Application])

Applies to Amazon EMR releases 4.0 and later. A case-insensitive list of applications for Amazon EMR to install and configure when launching the cluster. For a list of applications available for each Amazon EMR release version, see the Amazon EMR Release Guide.

runJobFlow_tags :: Lens' RunJobFlow (Maybe [Tag])

A list of tags to associate with a cluster and propagate to Amazon EC2 instances.

runJobFlow_serviceRole :: Lens' RunJobFlow (Maybe Text)

The IAM role that Amazon EMR assumes in order to access Amazon Web Services resources on your behalf.

runJobFlow_name :: Lens' RunJobFlow Text

The name of the job flow.

runJobFlow_instances :: Lens' RunJobFlow JobFlowInstancesConfig

A specification of the number and type of Amazon EC2 instances.

Destructuring the Response

data RunJobFlowResponse

The result of the RunJobFlow operation.

See: newRunJobFlowResponse smart constructor.

Constructors

RunJobFlowResponse'

Fields

clusterArn :: Maybe Text

The Amazon Resource Name (ARN) of the cluster.

jobFlowId :: Maybe Text

A unique identifier for the job flow.

httpStatus :: Int

The response's http status code.

newRunJobFlowResponse

Arguments

:: Int
$sel:httpStatus:RunJobFlowResponse'

-> RunJobFlowResponse

Create a value of RunJobFlowResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clusterArn:RunJobFlowResponse', runJobFlowResponse_clusterArn - The Amazon Resource Name (ARN) of the cluster.

$sel:jobFlowId:RunJobFlowResponse', runJobFlowResponse_jobFlowId - A unique identifier for the job flow.

$sel:httpStatus:RunJobFlowResponse', runJobFlowResponse_httpStatus - The response's http status code.

Response Lenses

runJobFlowResponse_clusterArn :: Lens' RunJobFlowResponse (Maybe Text)

The Amazon Resource Name (ARN) of the cluster.

runJobFlowResponse_jobFlowId :: Lens' RunJobFlowResponse (Maybe Text)

A unique identifier for the job flow.

runJobFlowResponse_httpStatus :: Lens' RunJobFlowResponse Int

The response's http status code.