Network.AWS.EMR.Types.CloudWatchAlarmDefinition

Description

Documentation

data CloudWatchAlarmDefinition

The definition of a CloudWatch metric alarm, which determines when an automatic scaling activity is triggered. When the defined alarm conditions are satisfied, scaling activity begins.

See: newCloudWatchAlarmDefinition smart constructor.

Constructors

CloudWatchAlarmDefinition'

Fields

unit :: Maybe Unit

The unit of measure associated with the CloudWatch metric being watched. The value specified for Unit must correspond to the units specified in the CloudWatch metric.

statistic :: Maybe Statistic

The statistic to apply to the metric associated with the alarm. The default is AVERAGE.

namespace :: Maybe Text

The namespace for the CloudWatch metric. The default is AWS/ElasticMapReduce.

dimensions :: Maybe [MetricDimension]

A CloudWatch metric dimension.

evaluationPeriods :: Maybe Int

The number of periods, in five-minute increments, during which the alarm condition must exist before the alarm triggers automatic scaling activity. The default value is 1.

comparisonOperator :: ComparisonOperator

Determines how the metric specified by MetricName is compared to the value specified by Threshold.

metricName :: Text

The name of the CloudWatch metric that is watched to determine an alarm condition.

period :: Int

The period, in seconds, over which the statistic is applied. EMR CloudWatch metrics are emitted every five minutes (300 seconds), so if an EMR CloudWatch metric is specified, specify 300.

threshold :: Double

The value against which the specified statistic is compared.

newCloudWatchAlarmDefinition

Arguments

:: ComparisonOperator
$sel:comparisonOperator:CloudWatchAlarmDefinition'

-> Text
$sel:metricName:CloudWatchAlarmDefinition'

-> Int
$sel:period:CloudWatchAlarmDefinition'

-> Double
$sel:threshold:CloudWatchAlarmDefinition'

-> CloudWatchAlarmDefinition

Create a value of CloudWatchAlarmDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:unit:CloudWatchAlarmDefinition', cloudWatchAlarmDefinition_unit - The unit of measure associated with the CloudWatch metric being watched. The value specified for Unit must correspond to the units specified in the CloudWatch metric.

$sel:statistic:CloudWatchAlarmDefinition', cloudWatchAlarmDefinition_statistic - The statistic to apply to the metric associated with the alarm. The default is AVERAGE.

$sel:namespace:CloudWatchAlarmDefinition', cloudWatchAlarmDefinition_namespace - The namespace for the CloudWatch metric. The default is AWS/ElasticMapReduce.

$sel:dimensions:CloudWatchAlarmDefinition', cloudWatchAlarmDefinition_dimensions - A CloudWatch metric dimension.

$sel:evaluationPeriods:CloudWatchAlarmDefinition', cloudWatchAlarmDefinition_evaluationPeriods - The number of periods, in five-minute increments, during which the alarm condition must exist before the alarm triggers automatic scaling activity. The default value is 1.

$sel:comparisonOperator:CloudWatchAlarmDefinition', cloudWatchAlarmDefinition_comparisonOperator - Determines how the metric specified by MetricName is compared to the value specified by Threshold.

$sel:metricName:CloudWatchAlarmDefinition', cloudWatchAlarmDefinition_metricName - The name of the CloudWatch metric that is watched to determine an alarm condition.

$sel:period:CloudWatchAlarmDefinition', cloudWatchAlarmDefinition_period - The period, in seconds, over which the statistic is applied. EMR CloudWatch metrics are emitted every five minutes (300 seconds), so if an EMR CloudWatch metric is specified, specify 300.

$sel:threshold:CloudWatchAlarmDefinition', cloudWatchAlarmDefinition_threshold - The value against which the specified statistic is compared.

cloudWatchAlarmDefinition_unit :: Lens' CloudWatchAlarmDefinition (Maybe Unit)

The unit of measure associated with the CloudWatch metric being watched. The value specified for Unit must correspond to the units specified in the CloudWatch metric.

cloudWatchAlarmDefinition_statistic :: Lens' CloudWatchAlarmDefinition (Maybe Statistic)

The statistic to apply to the metric associated with the alarm. The default is AVERAGE.

cloudWatchAlarmDefinition_namespace :: Lens' CloudWatchAlarmDefinition (Maybe Text)

The namespace for the CloudWatch metric. The default is AWS/ElasticMapReduce.

cloudWatchAlarmDefinition_dimensions :: Lens' CloudWatchAlarmDefinition (Maybe [MetricDimension])

A CloudWatch metric dimension.

cloudWatchAlarmDefinition_evaluationPeriods :: Lens' CloudWatchAlarmDefinition (Maybe Int)

The number of periods, in five-minute increments, during which the alarm condition must exist before the alarm triggers automatic scaling activity. The default value is 1.

cloudWatchAlarmDefinition_comparisonOperator :: Lens' CloudWatchAlarmDefinition ComparisonOperator

Determines how the metric specified by MetricName is compared to the value specified by Threshold.

cloudWatchAlarmDefinition_metricName :: Lens' CloudWatchAlarmDefinition Text

The name of the CloudWatch metric that is watched to determine an alarm condition.

cloudWatchAlarmDefinition_period :: Lens' CloudWatchAlarmDefinition Int

The period, in seconds, over which the statistic is applied. EMR CloudWatch metrics are emitted every five minutes (300 seconds), so if an EMR CloudWatch metric is specified, specify 300.

cloudWatchAlarmDefinition_threshold :: Lens' CloudWatchAlarmDefinition Double

The value against which the specified statistic is compared.