Network.AWS.EMR.Types.ClusterSummary

Description

Documentation

data ClusterSummary

The summary description of the cluster.

See: newClusterSummary smart constructor.

Constructors

ClusterSummary'

Fields

clusterArn :: Maybe Text

The Amazon Resource Name of the cluster.

status :: Maybe ClusterStatus

The details about the current status of the cluster.

outpostArn :: Maybe Text

The Amazon Resource Name (ARN) of the Outpost where the cluster is launched.

id :: Maybe Text

The unique identifier for the cluster.

name :: Maybe Text

The name of the cluster.

normalizedInstanceHours :: Maybe Int

An approximation of the cost of the cluster, represented in m1.small/hours. This value is incremented one time for every hour an m1.small instance runs. Larger instances are weighted more, so an EC2 instance that is roughly four times more expensive would result in the normalized instance hours being incremented by four. This result is only an approximation and does not reflect the actual billing rate.

newClusterSummary :: ClusterSummary

Create a value of ClusterSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clusterArn:ClusterSummary', clusterSummary_clusterArn - The Amazon Resource Name of the cluster.

$sel:status:ClusterSummary', clusterSummary_status - The details about the current status of the cluster.

$sel:outpostArn:ClusterSummary', clusterSummary_outpostArn - The Amazon Resource Name (ARN) of the Outpost where the cluster is launched.

$sel:id:ClusterSummary', clusterSummary_id - The unique identifier for the cluster.

$sel:name:ClusterSummary', clusterSummary_name - The name of the cluster.

$sel:normalizedInstanceHours:ClusterSummary', clusterSummary_normalizedInstanceHours - An approximation of the cost of the cluster, represented in m1.small/hours. This value is incremented one time for every hour an m1.small instance runs. Larger instances are weighted more, so an EC2 instance that is roughly four times more expensive would result in the normalized instance hours being incremented by four. This result is only an approximation and does not reflect the actual billing rate.

clusterSummary_clusterArn :: Lens' ClusterSummary (Maybe Text)

The Amazon Resource Name of the cluster.

clusterSummary_status :: Lens' ClusterSummary (Maybe ClusterStatus)

The details about the current status of the cluster.

clusterSummary_outpostArn :: Lens' ClusterSummary (Maybe Text)

The Amazon Resource Name (ARN) of the Outpost where the cluster is launched.

clusterSummary_id :: Lens' ClusterSummary (Maybe Text)

The unique identifier for the cluster.

clusterSummary_name :: Lens' ClusterSummary (Maybe Text)

The name of the cluster.

clusterSummary_normalizedInstanceHours :: Lens' ClusterSummary (Maybe Int)

An approximation of the cost of the cluster, represented in m1.small/hours. This value is incremented one time for every hour an m1.small instance runs. Larger instances are weighted more, so an EC2 instance that is roughly four times more expensive would result in the normalized instance hours being incremented by four. This result is only an approximation and does not reflect the actual billing rate.