Network.AWS.EMR.Types.ComputeLimits

Description

Documentation

data ComputeLimits

The EC2 unit limits for a managed scaling policy. The managed scaling activity of a cluster can not be above or below these limits. The limit only applies to the core and task nodes. The master node cannot be scaled after initial configuration.

See: newComputeLimits smart constructor.

Constructors

ComputeLimits'

Fields

maximumOnDemandCapacityUnits :: Maybe Int

The upper boundary of On-Demand EC2 units. It is measured through vCPU cores or instances for instance groups and measured through units for instance fleets. The On-Demand units are not allowed to scale beyond this boundary. The parameter is used to split capacity allocation between On-Demand and Spot Instances.

maximumCoreCapacityUnits :: Maybe Int

The upper boundary of EC2 units for core node type in a cluster. It is measured through vCPU cores or instances for instance groups and measured through units for instance fleets. The core units are not allowed to scale beyond this boundary. The parameter is used to split capacity allocation between core and task nodes.

unitType :: ComputeLimitsUnitType

The unit type used for specifying a managed scaling policy.

minimumCapacityUnits :: Int

The lower boundary of EC2 units. It is measured through vCPU cores or instances for instance groups and measured through units for instance fleets. Managed scaling activities are not allowed beyond this boundary. The limit only applies to the core and task nodes. The master node cannot be scaled after initial configuration.

maximumCapacityUnits :: Int

The upper boundary of EC2 units. It is measured through vCPU cores or instances for instance groups and measured through units for instance fleets. Managed scaling activities are not allowed beyond this boundary. The limit only applies to the core and task nodes. The master node cannot be scaled after initial configuration.

newComputeLimits

Arguments

:: ComputeLimitsUnitType
$sel:unitType:ComputeLimits'

-> Int
$sel:minimumCapacityUnits:ComputeLimits'

-> Int
$sel:maximumCapacityUnits:ComputeLimits'

-> ComputeLimits

Create a value of ComputeLimits with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maximumOnDemandCapacityUnits:ComputeLimits', computeLimits_maximumOnDemandCapacityUnits - The upper boundary of On-Demand EC2 units. It is measured through vCPU cores or instances for instance groups and measured through units for instance fleets. The On-Demand units are not allowed to scale beyond this boundary. The parameter is used to split capacity allocation between On-Demand and Spot Instances.

$sel:maximumCoreCapacityUnits:ComputeLimits', computeLimits_maximumCoreCapacityUnits - The upper boundary of EC2 units for core node type in a cluster. It is measured through vCPU cores or instances for instance groups and measured through units for instance fleets. The core units are not allowed to scale beyond this boundary. The parameter is used to split capacity allocation between core and task nodes.

$sel:unitType:ComputeLimits', computeLimits_unitType - The unit type used for specifying a managed scaling policy.

$sel:minimumCapacityUnits:ComputeLimits', computeLimits_minimumCapacityUnits - The lower boundary of EC2 units. It is measured through vCPU cores or instances for instance groups and measured through units for instance fleets. Managed scaling activities are not allowed beyond this boundary. The limit only applies to the core and task nodes. The master node cannot be scaled after initial configuration.

$sel:maximumCapacityUnits:ComputeLimits', computeLimits_maximumCapacityUnits - The upper boundary of EC2 units. It is measured through vCPU cores or instances for instance groups and measured through units for instance fleets. Managed scaling activities are not allowed beyond this boundary. The limit only applies to the core and task nodes. The master node cannot be scaled after initial configuration.

computeLimits_maximumOnDemandCapacityUnits :: Lens' ComputeLimits (Maybe Int)

The upper boundary of On-Demand EC2 units. It is measured through vCPU cores or instances for instance groups and measured through units for instance fleets. The On-Demand units are not allowed to scale beyond this boundary. The parameter is used to split capacity allocation between On-Demand and Spot Instances.

computeLimits_maximumCoreCapacityUnits :: Lens' ComputeLimits (Maybe Int)

The upper boundary of EC2 units for core node type in a cluster. It is measured through vCPU cores or instances for instance groups and measured through units for instance fleets. The core units are not allowed to scale beyond this boundary. The parameter is used to split capacity allocation between core and task nodes.

computeLimits_unitType :: Lens' ComputeLimits ComputeLimitsUnitType

The unit type used for specifying a managed scaling policy.

computeLimits_minimumCapacityUnits :: Lens' ComputeLimits Int

The lower boundary of EC2 units. It is measured through vCPU cores or instances for instance groups and measured through units for instance fleets. Managed scaling activities are not allowed beyond this boundary. The limit only applies to the core and task nodes. The master node cannot be scaled after initial configuration.

computeLimits_maximumCapacityUnits :: Lens' ComputeLimits Int

The upper boundary of EC2 units. It is measured through vCPU cores or instances for instance groups and measured through units for instance fleets. Managed scaling activities are not allowed beyond this boundary. The limit only applies to the core and task nodes. The master node cannot be scaled after initial configuration.