Network.AWS.EMR.Types.Ec2InstanceAttributes

Description

Documentation

data Ec2InstanceAttributes

Provides information about the EC2 instances in a cluster grouped by category. For example, key name, subnet ID, IAM instance profile, and so on.

See: newEc2InstanceAttributes smart constructor.

Constructors

Ec2InstanceAttributes'

Fields

ec2KeyName :: Maybe Text

The name of the Amazon EC2 key pair to use when connecting with SSH into the master node as a user named "hadoop".

emrManagedSlaveSecurityGroup :: Maybe Text

The identifier of the Amazon EC2 security group for the core and task nodes.

additionalSlaveSecurityGroups :: Maybe [Text]

A list of additional Amazon EC2 security group IDs for the core and task nodes.

requestedEc2SubnetIds :: Maybe [Text]

Applies to clusters configured with the instance fleets option. Specifies the unique identifier of one or more Amazon EC2 subnets in which to launch EC2 cluster instances. Subnets must exist within the same VPC. Amazon EMR chooses the EC2 subnet with the best fit from among the list of RequestedEc2SubnetIds, and then launches all cluster instances within that Subnet. If this value is not specified, and the account and Region support EC2-Classic networks, the cluster launches instances in the EC2-Classic network and uses RequestedEc2AvailabilityZones instead of this setting. If EC2-Classic is not supported, and no Subnet is specified, Amazon EMR chooses the subnet for you. RequestedEc2SubnetIDs and RequestedEc2AvailabilityZones cannot be specified together.

additionalMasterSecurityGroups :: Maybe [Text]

A list of additional Amazon EC2 security group IDs for the master node.

iamInstanceProfile :: Maybe Text

The IAM role that was specified when the cluster was launched. The EC2 instances of the cluster assume this role.

emrManagedMasterSecurityGroup :: Maybe Text

The identifier of the Amazon EC2 security group for the master node.

ec2SubnetId :: Maybe Text

Set this parameter to the identifier of the Amazon VPC subnet where you want the cluster to launch. If you do not specify this value, and your account supports EC2-Classic, the cluster launches in EC2-Classic.

requestedEc2AvailabilityZones :: Maybe [Text]

Applies to clusters configured with the instance fleets option. Specifies one or more Availability Zones in which to launch EC2 cluster instances when the EC2-Classic network configuration is supported. Amazon EMR chooses the Availability Zone with the best fit from among the list of RequestedEc2AvailabilityZones, and then launches all cluster instances within that Availability Zone. If you do not specify this value, Amazon EMR chooses the Availability Zone for you. RequestedEc2SubnetIDs and RequestedEc2AvailabilityZones cannot be specified together.

serviceAccessSecurityGroup :: Maybe Text

The identifier of the Amazon EC2 security group for the Amazon EMR service to access clusters in VPC private subnets.

ec2AvailabilityZone :: Maybe Text

The Availability Zone in which the cluster will run.

newEc2InstanceAttributes :: Ec2InstanceAttributes

Create a value of Ec2InstanceAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ec2KeyName:Ec2InstanceAttributes', ec2InstanceAttributes_ec2KeyName - The name of the Amazon EC2 key pair to use when connecting with SSH into the master node as a user named "hadoop".

$sel:emrManagedSlaveSecurityGroup:Ec2InstanceAttributes', ec2InstanceAttributes_emrManagedSlaveSecurityGroup - The identifier of the Amazon EC2 security group for the core and task nodes.

$sel:additionalSlaveSecurityGroups:Ec2InstanceAttributes', ec2InstanceAttributes_additionalSlaveSecurityGroups - A list of additional Amazon EC2 security group IDs for the core and task nodes.

$sel:requestedEc2SubnetIds:Ec2InstanceAttributes', ec2InstanceAttributes_requestedEc2SubnetIds - Applies to clusters configured with the instance fleets option. Specifies the unique identifier of one or more Amazon EC2 subnets in which to launch EC2 cluster instances. Subnets must exist within the same VPC. Amazon EMR chooses the EC2 subnet with the best fit from among the list of RequestedEc2SubnetIds, and then launches all cluster instances within that Subnet. If this value is not specified, and the account and Region support EC2-Classic networks, the cluster launches instances in the EC2-Classic network and uses RequestedEc2AvailabilityZones instead of this setting. If EC2-Classic is not supported, and no Subnet is specified, Amazon EMR chooses the subnet for you. RequestedEc2SubnetIDs and RequestedEc2AvailabilityZones cannot be specified together.

$sel:additionalMasterSecurityGroups:Ec2InstanceAttributes', ec2InstanceAttributes_additionalMasterSecurityGroups - A list of additional Amazon EC2 security group IDs for the master node.

$sel:iamInstanceProfile:Ec2InstanceAttributes', ec2InstanceAttributes_iamInstanceProfile - The IAM role that was specified when the cluster was launched. The EC2 instances of the cluster assume this role.

$sel:emrManagedMasterSecurityGroup:Ec2InstanceAttributes', ec2InstanceAttributes_emrManagedMasterSecurityGroup - The identifier of the Amazon EC2 security group for the master node.

$sel:ec2SubnetId:Ec2InstanceAttributes', ec2InstanceAttributes_ec2SubnetId - Set this parameter to the identifier of the Amazon VPC subnet where you want the cluster to launch. If you do not specify this value, and your account supports EC2-Classic, the cluster launches in EC2-Classic.

$sel:requestedEc2AvailabilityZones:Ec2InstanceAttributes', ec2InstanceAttributes_requestedEc2AvailabilityZones - Applies to clusters configured with the instance fleets option. Specifies one or more Availability Zones in which to launch EC2 cluster instances when the EC2-Classic network configuration is supported. Amazon EMR chooses the Availability Zone with the best fit from among the list of RequestedEc2AvailabilityZones, and then launches all cluster instances within that Availability Zone. If you do not specify this value, Amazon EMR chooses the Availability Zone for you. RequestedEc2SubnetIDs and RequestedEc2AvailabilityZones cannot be specified together.

$sel:serviceAccessSecurityGroup:Ec2InstanceAttributes', ec2InstanceAttributes_serviceAccessSecurityGroup - The identifier of the Amazon EC2 security group for the Amazon EMR service to access clusters in VPC private subnets.

$sel:ec2AvailabilityZone:Ec2InstanceAttributes', ec2InstanceAttributes_ec2AvailabilityZone - The Availability Zone in which the cluster will run.

ec2InstanceAttributes_ec2KeyName :: Lens' Ec2InstanceAttributes (Maybe Text)

The name of the Amazon EC2 key pair to use when connecting with SSH into the master node as a user named "hadoop".

ec2InstanceAttributes_emrManagedSlaveSecurityGroup :: Lens' Ec2InstanceAttributes (Maybe Text)

The identifier of the Amazon EC2 security group for the core and task nodes.

ec2InstanceAttributes_additionalSlaveSecurityGroups :: Lens' Ec2InstanceAttributes (Maybe [Text])

A list of additional Amazon EC2 security group IDs for the core and task nodes.

ec2InstanceAttributes_requestedEc2SubnetIds :: Lens' Ec2InstanceAttributes (Maybe [Text])

Applies to clusters configured with the instance fleets option. Specifies the unique identifier of one or more Amazon EC2 subnets in which to launch EC2 cluster instances. Subnets must exist within the same VPC. Amazon EMR chooses the EC2 subnet with the best fit from among the list of RequestedEc2SubnetIds, and then launches all cluster instances within that Subnet. If this value is not specified, and the account and Region support EC2-Classic networks, the cluster launches instances in the EC2-Classic network and uses RequestedEc2AvailabilityZones instead of this setting. If EC2-Classic is not supported, and no Subnet is specified, Amazon EMR chooses the subnet for you. RequestedEc2SubnetIDs and RequestedEc2AvailabilityZones cannot be specified together.

ec2InstanceAttributes_additionalMasterSecurityGroups :: Lens' Ec2InstanceAttributes (Maybe [Text])

A list of additional Amazon EC2 security group IDs for the master node.

ec2InstanceAttributes_iamInstanceProfile :: Lens' Ec2InstanceAttributes (Maybe Text)

The IAM role that was specified when the cluster was launched. The EC2 instances of the cluster assume this role.

ec2InstanceAttributes_emrManagedMasterSecurityGroup :: Lens' Ec2InstanceAttributes (Maybe Text)

The identifier of the Amazon EC2 security group for the master node.

ec2InstanceAttributes_ec2SubnetId :: Lens' Ec2InstanceAttributes (Maybe Text)

Set this parameter to the identifier of the Amazon VPC subnet where you want the cluster to launch. If you do not specify this value, and your account supports EC2-Classic, the cluster launches in EC2-Classic.

ec2InstanceAttributes_requestedEc2AvailabilityZones :: Lens' Ec2InstanceAttributes (Maybe [Text])

Applies to clusters configured with the instance fleets option. Specifies one or more Availability Zones in which to launch EC2 cluster instances when the EC2-Classic network configuration is supported. Amazon EMR chooses the Availability Zone with the best fit from among the list of RequestedEc2AvailabilityZones, and then launches all cluster instances within that Availability Zone. If you do not specify this value, Amazon EMR chooses the Availability Zone for you. RequestedEc2SubnetIDs and RequestedEc2AvailabilityZones cannot be specified together.

ec2InstanceAttributes_serviceAccessSecurityGroup :: Lens' Ec2InstanceAttributes (Maybe Text)

The identifier of the Amazon EC2 security group for the Amazon EMR service to access clusters in VPC private subnets.

ec2InstanceAttributes_ec2AvailabilityZone :: Lens' Ec2InstanceAttributes (Maybe Text)

The Availability Zone in which the cluster will run.