Network.AWS.EMR.Types.InstanceFleetStatus

Description

Documentation

data InstanceFleetStatus

The status of the instance fleet.

The instance fleet configuration is available only in Amazon EMR versions 4.8.0 and later, excluding 5.0.x versions.

See: newInstanceFleetStatus smart constructor.

Constructors

InstanceFleetStatus'

Fields

stateChangeReason :: Maybe InstanceFleetStateChangeReason

Provides status change reason details for the instance fleet.

state :: Maybe InstanceFleetState

A code representing the instance fleet status.

PROVISIONING—The instance fleet is provisioning EC2 resources and is not yet ready to run jobs.

BOOTSTRAPPING—EC2 instances and other resources have been provisioned and the bootstrap actions specified for the instances are underway.

RUNNING—EC2 instances and other resources are running. They are either executing jobs or waiting to execute jobs.

RESIZING—A resize operation is underway. EC2 instances are either being added or removed.

SUSPENDED—A resize operation could not complete. Existing EC2 instances are running, but instances can't be added or removed.

TERMINATING—The instance fleet is terminating EC2 instances.

TERMINATED—The instance fleet is no longer active, and all EC2 instances have been terminated.

timeline :: Maybe InstanceFleetTimeline

Provides historical timestamps for the instance fleet, including the time of creation, the time it became ready to run jobs, and the time of termination.

newInstanceFleetStatus :: InstanceFleetStatus

Create a value of InstanceFleetStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stateChangeReason:InstanceFleetStatus', instanceFleetStatus_stateChangeReason - Provides status change reason details for the instance fleet.

$sel:state:InstanceFleetStatus', instanceFleetStatus_state - A code representing the instance fleet status.

PROVISIONING—The instance fleet is provisioning EC2 resources and is not yet ready to run jobs.

BOOTSTRAPPING—EC2 instances and other resources have been provisioned and the bootstrap actions specified for the instances are underway.

RUNNING—EC2 instances and other resources are running. They are either executing jobs or waiting to execute jobs.

RESIZING—A resize operation is underway. EC2 instances are either being added or removed.

SUSPENDED—A resize operation could not complete. Existing EC2 instances are running, but instances can't be added or removed.

TERMINATING—The instance fleet is terminating EC2 instances.

TERMINATED—The instance fleet is no longer active, and all EC2 instances have been terminated.

$sel:timeline:InstanceFleetStatus', instanceFleetStatus_timeline - Provides historical timestamps for the instance fleet, including the time of creation, the time it became ready to run jobs, and the time of termination.

instanceFleetStatus_stateChangeReason :: Lens' InstanceFleetStatus (Maybe InstanceFleetStateChangeReason)

Provides status change reason details for the instance fleet.

instanceFleetStatus_state :: Lens' InstanceFleetStatus (Maybe InstanceFleetState)

A code representing the instance fleet status.

PROVISIONING—The instance fleet is provisioning EC2 resources and is not yet ready to run jobs.

BOOTSTRAPPING—EC2 instances and other resources have been provisioned and the bootstrap actions specified for the instances are underway.

RUNNING—EC2 instances and other resources are running. They are either executing jobs or waiting to execute jobs.

RESIZING—A resize operation is underway. EC2 instances are either being added or removed.

SUSPENDED—A resize operation could not complete. Existing EC2 instances are running, but instances can't be added or removed.

TERMINATING—The instance fleet is terminating EC2 instances.

TERMINATED—The instance fleet is no longer active, and all EC2 instances have been terminated.

instanceFleetStatus_timeline :: Lens' InstanceFleetStatus (Maybe InstanceFleetTimeline)

Provides historical timestamps for the instance fleet, including the time of creation, the time it became ready to run jobs, and the time of termination.