Network.AWS.EMR.Types.InstanceGroup

Description

Documentation

data InstanceGroup

This entity represents an instance group, which is a group of instances that have common purpose. For example, CORE instance group is used for HDFS.

See: newInstanceGroup smart constructor.

Constructors

InstanceGroup'

Fields

lastSuccessfullyAppliedConfigurationsVersion :: Maybe Integer

The version number of a configuration specification that was successfully applied for an instance group last time.

status :: Maybe InstanceGroupStatus

The current status of the instance group.

ebsOptimized :: Maybe Bool

If the instance group is EBS-optimized. An Amazon EBS-optimized instance uses an optimized configuration stack and provides additional, dedicated capacity for Amazon EBS I/O.

instanceType :: Maybe Text

The EC2 instance type for all instances in the instance group.

ebsBlockDevices :: Maybe [EbsBlockDevice]

The EBS block devices that are mapped to this instance group.

instanceGroupType :: Maybe InstanceGroupType

The type of the instance group. Valid values are MASTER, CORE or TASK.

configurations :: Maybe [Configuration]

Amazon EMR releases 4.x or later.

The list of configurations supplied for an Amazon EMR cluster instance group. You can specify a separate configuration for each instance group (master, core, and task).

id :: Maybe Text

The identifier of the instance group.

lastSuccessfullyAppliedConfigurations :: Maybe [Configuration]

A list of configurations that were successfully applied for an instance group last time.

shrinkPolicy :: Maybe ShrinkPolicy

Policy for customizing shrink operations.

bidPrice :: Maybe Text

If specified, indicates that the instance group uses Spot Instances. This is the maximum price you are willing to pay for Spot Instances. Specify OnDemandPrice to set the amount equal to the On-Demand price, or specify an amount in USD.

requestedInstanceCount :: Maybe Int

The target number of instances for the instance group.

autoScalingPolicy :: Maybe AutoScalingPolicyDescription

An automatic scaling policy for a core instance group or task instance group in an Amazon EMR cluster. The automatic scaling policy defines how an instance group dynamically adds and terminates EC2 instances in response to the value of a CloudWatch metric. See PutAutoScalingPolicy.

name :: Maybe Text

The name of the instance group.

market :: Maybe MarketType

The marketplace to provision instances for this group. Valid values are ON_DEMAND or SPOT.

configurationsVersion :: Maybe Integer

The version number of the requested configuration specification for this instance group.

customAmiId :: Maybe Text

The custom AMI ID to use for the provisioned instance group.

runningInstanceCount :: Maybe Int

The number of instances currently running in this instance group.

newInstanceGroup :: InstanceGroup

Create a value of InstanceGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastSuccessfullyAppliedConfigurationsVersion:InstanceGroup', instanceGroup_lastSuccessfullyAppliedConfigurationsVersion - The version number of a configuration specification that was successfully applied for an instance group last time.

$sel:status:InstanceGroup', instanceGroup_status - The current status of the instance group.

$sel:ebsOptimized:InstanceGroup', instanceGroup_ebsOptimized - If the instance group is EBS-optimized. An Amazon EBS-optimized instance uses an optimized configuration stack and provides additional, dedicated capacity for Amazon EBS I/O.

$sel:instanceType:InstanceGroup', instanceGroup_instanceType - The EC2 instance type for all instances in the instance group.

$sel:ebsBlockDevices:InstanceGroup', instanceGroup_ebsBlockDevices - The EBS block devices that are mapped to this instance group.

$sel:instanceGroupType:InstanceGroup', instanceGroup_instanceGroupType - The type of the instance group. Valid values are MASTER, CORE or TASK.

$sel:configurations:InstanceGroup', instanceGroup_configurations - Amazon EMR releases 4.x or later.

The list of configurations supplied for an Amazon EMR cluster instance group. You can specify a separate configuration for each instance group (master, core, and task).

$sel:id:InstanceGroup', instanceGroup_id - The identifier of the instance group.

$sel:lastSuccessfullyAppliedConfigurations:InstanceGroup', instanceGroup_lastSuccessfullyAppliedConfigurations - A list of configurations that were successfully applied for an instance group last time.

$sel:shrinkPolicy:InstanceGroup', instanceGroup_shrinkPolicy - Policy for customizing shrink operations.

$sel:bidPrice:InstanceGroup', instanceGroup_bidPrice - If specified, indicates that the instance group uses Spot Instances. This is the maximum price you are willing to pay for Spot Instances. Specify OnDemandPrice to set the amount equal to the On-Demand price, or specify an amount in USD.

$sel:requestedInstanceCount:InstanceGroup', instanceGroup_requestedInstanceCount - The target number of instances for the instance group.

$sel:autoScalingPolicy:InstanceGroup', instanceGroup_autoScalingPolicy - An automatic scaling policy for a core instance group or task instance group in an Amazon EMR cluster. The automatic scaling policy defines how an instance group dynamically adds and terminates EC2 instances in response to the value of a CloudWatch metric. See PutAutoScalingPolicy.

$sel:name:InstanceGroup', instanceGroup_name - The name of the instance group.

$sel:market:InstanceGroup', instanceGroup_market - The marketplace to provision instances for this group. Valid values are ON_DEMAND or SPOT.

$sel:configurationsVersion:InstanceGroup', instanceGroup_configurationsVersion - The version number of the requested configuration specification for this instance group.

$sel:customAmiId:InstanceGroup', instanceGroup_customAmiId - The custom AMI ID to use for the provisioned instance group.

$sel:runningInstanceCount:InstanceGroup', instanceGroup_runningInstanceCount - The number of instances currently running in this instance group.

instanceGroup_lastSuccessfullyAppliedConfigurationsVersion :: Lens' InstanceGroup (Maybe Integer)

The version number of a configuration specification that was successfully applied for an instance group last time.

instanceGroup_status :: Lens' InstanceGroup (Maybe InstanceGroupStatus)

The current status of the instance group.

instanceGroup_ebsOptimized :: Lens' InstanceGroup (Maybe Bool)

If the instance group is EBS-optimized. An Amazon EBS-optimized instance uses an optimized configuration stack and provides additional, dedicated capacity for Amazon EBS I/O.

instanceGroup_instanceType :: Lens' InstanceGroup (Maybe Text)

The EC2 instance type for all instances in the instance group.

instanceGroup_ebsBlockDevices :: Lens' InstanceGroup (Maybe [EbsBlockDevice])

The EBS block devices that are mapped to this instance group.

instanceGroup_instanceGroupType :: Lens' InstanceGroup (Maybe InstanceGroupType)

The type of the instance group. Valid values are MASTER, CORE or TASK.

instanceGroup_configurations :: Lens' InstanceGroup (Maybe [Configuration])

Amazon EMR releases 4.x or later.

The list of configurations supplied for an Amazon EMR cluster instance group. You can specify a separate configuration for each instance group (master, core, and task).

instanceGroup_id :: Lens' InstanceGroup (Maybe Text)

The identifier of the instance group.

instanceGroup_lastSuccessfullyAppliedConfigurations :: Lens' InstanceGroup (Maybe [Configuration])

A list of configurations that were successfully applied for an instance group last time.

instanceGroup_shrinkPolicy :: Lens' InstanceGroup (Maybe ShrinkPolicy)

Policy for customizing shrink operations.

instanceGroup_bidPrice :: Lens' InstanceGroup (Maybe Text)

If specified, indicates that the instance group uses Spot Instances. This is the maximum price you are willing to pay for Spot Instances. Specify OnDemandPrice to set the amount equal to the On-Demand price, or specify an amount in USD.

instanceGroup_requestedInstanceCount :: Lens' InstanceGroup (Maybe Int)

The target number of instances for the instance group.

instanceGroup_autoScalingPolicy :: Lens' InstanceGroup (Maybe AutoScalingPolicyDescription)

An automatic scaling policy for a core instance group or task instance group in an Amazon EMR cluster. The automatic scaling policy defines how an instance group dynamically adds and terminates EC2 instances in response to the value of a CloudWatch metric. See PutAutoScalingPolicy.

instanceGroup_name :: Lens' InstanceGroup (Maybe Text)

The name of the instance group.

instanceGroup_market :: Lens' InstanceGroup (Maybe MarketType)

The marketplace to provision instances for this group. Valid values are ON_DEMAND or SPOT.

instanceGroup_configurationsVersion :: Lens' InstanceGroup (Maybe Integer)

The version number of the requested configuration specification for this instance group.

instanceGroup_customAmiId :: Lens' InstanceGroup (Maybe Text)

The custom AMI ID to use for the provisioned instance group.

instanceGroup_runningInstanceCount :: Lens' InstanceGroup (Maybe Int)

The number of instances currently running in this instance group.