Network.AWS.EMR.Types.InstanceGroupConfig

Description

Documentation

data InstanceGroupConfig

Configuration defining a new instance group.

See: newInstanceGroupConfig smart constructor.

Constructors

InstanceGroupConfig'

Fields

ebsConfiguration :: Maybe EbsConfiguration

EBS configurations that will be attached to each EC2 instance in the instance group.

configurations :: Maybe [Configuration]

Amazon EMR releases 4.x or later.

The list of configurations supplied for an EMR cluster instance group. You can specify a separate configuration for each instance group (master, core, and task).

bidPrice :: Maybe Text

If specified, indicates that the instance group uses Spot Instances. This is the maximum price you are willing to pay for Spot Instances. Specify OnDemandPrice to set the amount equal to the On-Demand price, or specify an amount in USD.

autoScalingPolicy :: Maybe AutoScalingPolicy

An automatic scaling policy for a core instance group or task instance group in an Amazon EMR cluster. The automatic scaling policy defines how an instance group dynamically adds and terminates EC2 instances in response to the value of a CloudWatch metric. See PutAutoScalingPolicy.

name :: Maybe Text

Friendly name given to the instance group.

market :: Maybe MarketType

Market type of the EC2 instances used to create a cluster node.

customAmiId :: Maybe Text

The custom AMI ID to use for the provisioned instance group.

instanceRole :: InstanceRoleType

The role of the instance group in the cluster.

instanceType :: Text

The EC2 instance type for all instances in the instance group.

instanceCount :: Int

Target number of instances for the instance group.

newInstanceGroupConfig

Arguments

:: InstanceRoleType
$sel:instanceRole:InstanceGroupConfig'

-> Text
$sel:instanceType:InstanceGroupConfig'

-> Int
$sel:instanceCount:InstanceGroupConfig'

-> InstanceGroupConfig

Create a value of InstanceGroupConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ebsConfiguration:InstanceGroupConfig', instanceGroupConfig_ebsConfiguration - EBS configurations that will be attached to each EC2 instance in the instance group.

$sel:configurations:InstanceGroupConfig', instanceGroupConfig_configurations - Amazon EMR releases 4.x or later.

The list of configurations supplied for an EMR cluster instance group. You can specify a separate configuration for each instance group (master, core, and task).

$sel:bidPrice:InstanceGroupConfig', instanceGroupConfig_bidPrice - If specified, indicates that the instance group uses Spot Instances. This is the maximum price you are willing to pay for Spot Instances. Specify OnDemandPrice to set the amount equal to the On-Demand price, or specify an amount in USD.

$sel:autoScalingPolicy:InstanceGroupConfig', instanceGroupConfig_autoScalingPolicy - An automatic scaling policy for a core instance group or task instance group in an Amazon EMR cluster. The automatic scaling policy defines how an instance group dynamically adds and terminates EC2 instances in response to the value of a CloudWatch metric. See PutAutoScalingPolicy.

$sel:name:InstanceGroupConfig', instanceGroupConfig_name - Friendly name given to the instance group.

$sel:market:InstanceGroupConfig', instanceGroupConfig_market - Market type of the EC2 instances used to create a cluster node.

$sel:customAmiId:InstanceGroupConfig', instanceGroupConfig_customAmiId - The custom AMI ID to use for the provisioned instance group.

$sel:instanceRole:InstanceGroupConfig', instanceGroupConfig_instanceRole - The role of the instance group in the cluster.

$sel:instanceType:InstanceGroupConfig', instanceGroupConfig_instanceType - The EC2 instance type for all instances in the instance group.

$sel:instanceCount:InstanceGroupConfig', instanceGroupConfig_instanceCount - Target number of instances for the instance group.

instanceGroupConfig_ebsConfiguration :: Lens' InstanceGroupConfig (Maybe EbsConfiguration)

EBS configurations that will be attached to each EC2 instance in the instance group.

instanceGroupConfig_configurations :: Lens' InstanceGroupConfig (Maybe [Configuration])

Amazon EMR releases 4.x or later.

The list of configurations supplied for an EMR cluster instance group. You can specify a separate configuration for each instance group (master, core, and task).

instanceGroupConfig_bidPrice :: Lens' InstanceGroupConfig (Maybe Text)

If specified, indicates that the instance group uses Spot Instances. This is the maximum price you are willing to pay for Spot Instances. Specify OnDemandPrice to set the amount equal to the On-Demand price, or specify an amount in USD.

instanceGroupConfig_autoScalingPolicy :: Lens' InstanceGroupConfig (Maybe AutoScalingPolicy)

An automatic scaling policy for a core instance group or task instance group in an Amazon EMR cluster. The automatic scaling policy defines how an instance group dynamically adds and terminates EC2 instances in response to the value of a CloudWatch metric. See PutAutoScalingPolicy.

instanceGroupConfig_name :: Lens' InstanceGroupConfig (Maybe Text)

Friendly name given to the instance group.

instanceGroupConfig_market :: Lens' InstanceGroupConfig (Maybe MarketType)

Market type of the EC2 instances used to create a cluster node.

instanceGroupConfig_customAmiId :: Lens' InstanceGroupConfig (Maybe Text)

The custom AMI ID to use for the provisioned instance group.

instanceGroupConfig_instanceRole :: Lens' InstanceGroupConfig InstanceRoleType

The role of the instance group in the cluster.

instanceGroupConfig_instanceType :: Lens' InstanceGroupConfig Text

The EC2 instance type for all instances in the instance group.

instanceGroupConfig_instanceCount :: Lens' InstanceGroupConfig Int

Target number of instances for the instance group.