Network.AWS.EMR.Types.InstanceTypeSpecification

Description

Documentation

data InstanceTypeSpecification

The configuration specification for each instance type in an instance fleet.

The instance fleet configuration is available only in Amazon EMR versions 4.8.0 and later, excluding 5.0.x versions.

See: newInstanceTypeSpecification smart constructor.

Constructors

InstanceTypeSpecification'

Fields

ebsOptimized :: Maybe Bool

Evaluates to TRUE when the specified InstanceType is EBS-optimized.

instanceType :: Maybe Text

The EC2 instance type, for example m3.xlarge.

ebsBlockDevices :: Maybe [EbsBlockDevice]

The configuration of Amazon Elastic Block Store (Amazon EBS) attached to each instance as defined by InstanceType.

configurations :: Maybe [Configuration]

A configuration classification that applies when provisioning cluster instances, which can include configurations for applications and software bundled with Amazon EMR.

bidPrice :: Maybe Text

The bid price for each EC2 Spot Instance type as defined by InstanceType. Expressed in USD.

bidPriceAsPercentageOfOnDemandPrice :: Maybe Double

The bid price, as a percentage of On-Demand price, for each EC2 Spot Instance as defined by InstanceType. Expressed as a number (for example, 20 specifies 20%).

customAmiId :: Maybe Text

The custom AMI ID to use for the instance type.

weightedCapacity :: Maybe Natural

The number of units that a provisioned instance of this type provides toward fulfilling the target capacities defined in InstanceFleetConfig. Capacity values represent performance characteristics such as vCPUs, memory, or I/O. If not specified, the default value is 1.

newInstanceTypeSpecification :: InstanceTypeSpecification

Create a value of InstanceTypeSpecification with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ebsOptimized:InstanceTypeSpecification', instanceTypeSpecification_ebsOptimized - Evaluates to TRUE when the specified InstanceType is EBS-optimized.

$sel:instanceType:InstanceTypeSpecification', instanceTypeSpecification_instanceType - The EC2 instance type, for example m3.xlarge.

$sel:ebsBlockDevices:InstanceTypeSpecification', instanceTypeSpecification_ebsBlockDevices - The configuration of Amazon Elastic Block Store (Amazon EBS) attached to each instance as defined by InstanceType.

$sel:configurations:InstanceTypeSpecification', instanceTypeSpecification_configurations - A configuration classification that applies when provisioning cluster instances, which can include configurations for applications and software bundled with Amazon EMR.

$sel:bidPrice:InstanceTypeSpecification', instanceTypeSpecification_bidPrice - The bid price for each EC2 Spot Instance type as defined by InstanceType. Expressed in USD.

$sel:bidPriceAsPercentageOfOnDemandPrice:InstanceTypeSpecification', instanceTypeSpecification_bidPriceAsPercentageOfOnDemandPrice - The bid price, as a percentage of On-Demand price, for each EC2 Spot Instance as defined by InstanceType. Expressed as a number (for example, 20 specifies 20%).

$sel:customAmiId:InstanceTypeSpecification', instanceTypeSpecification_customAmiId - The custom AMI ID to use for the instance type.

$sel:weightedCapacity:InstanceTypeSpecification', instanceTypeSpecification_weightedCapacity - The number of units that a provisioned instance of this type provides toward fulfilling the target capacities defined in InstanceFleetConfig. Capacity values represent performance characteristics such as vCPUs, memory, or I/O. If not specified, the default value is 1.

instanceTypeSpecification_ebsOptimized :: Lens' InstanceTypeSpecification (Maybe Bool)

Evaluates to TRUE when the specified InstanceType is EBS-optimized.

instanceTypeSpecification_instanceType :: Lens' InstanceTypeSpecification (Maybe Text)

The EC2 instance type, for example m3.xlarge.

instanceTypeSpecification_ebsBlockDevices :: Lens' InstanceTypeSpecification (Maybe [EbsBlockDevice])

The configuration of Amazon Elastic Block Store (Amazon EBS) attached to each instance as defined by InstanceType.

instanceTypeSpecification_configurations :: Lens' InstanceTypeSpecification (Maybe [Configuration])

A configuration classification that applies when provisioning cluster instances, which can include configurations for applications and software bundled with Amazon EMR.

instanceTypeSpecification_bidPrice :: Lens' InstanceTypeSpecification (Maybe Text)

The bid price for each EC2 Spot Instance type as defined by InstanceType. Expressed in USD.

instanceTypeSpecification_bidPriceAsPercentageOfOnDemandPrice :: Lens' InstanceTypeSpecification (Maybe Double)

The bid price, as a percentage of On-Demand price, for each EC2 Spot Instance as defined by InstanceType. Expressed as a number (for example, 20 specifies 20%).

instanceTypeSpecification_customAmiId :: Lens' InstanceTypeSpecification (Maybe Text)

The custom AMI ID to use for the instance type.

instanceTypeSpecification_weightedCapacity :: Lens' InstanceTypeSpecification (Maybe Natural)

The number of units that a provisioned instance of this type provides toward fulfilling the target capacities defined in InstanceFleetConfig. Capacity values represent performance characteristics such as vCPUs, memory, or I/O. If not specified, the default value is 1.