Network.AWS.EMR.Types.KerberosAttributes

Description

Documentation

data KerberosAttributes

Attributes for Kerberos configuration when Kerberos authentication is enabled using a security configuration. For more information see Use Kerberos Authentication in the Amazon EMR Management Guide.

See: newKerberosAttributes smart constructor.

Constructors

KerberosAttributes'

Fields

kdcAdminPassword :: Maybe Text

The password used within the cluster for the kadmin service on the cluster-dedicated KDC, which maintains Kerberos principals, password policies, and keytabs for the cluster.

realm :: Maybe Text

The name of the Kerberos realm to which all nodes in a cluster belong. For example, EC2.INTERNAL.

aDDomainJoinPassword :: Maybe Text

The Active Directory password for ADDomainJoinUser.

crossRealmTrustPrincipalPassword :: Maybe Text

Required only when establishing a cross-realm trust with a KDC in a different realm. The cross-realm principal password, which must be identical across realms.

aDDomainJoinUser :: Maybe Text

Required only when establishing a cross-realm trust with an Active Directory domain. A user with sufficient privileges to join resources to the domain.

newKerberosAttributes :: KerberosAttributes

Create a value of KerberosAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:kdcAdminPassword:KerberosAttributes', kerberosAttributes_kdcAdminPassword - The password used within the cluster for the kadmin service on the cluster-dedicated KDC, which maintains Kerberos principals, password policies, and keytabs for the cluster.

$sel:realm:KerberosAttributes', kerberosAttributes_realm - The name of the Kerberos realm to which all nodes in a cluster belong. For example, EC2.INTERNAL.

$sel:aDDomainJoinPassword:KerberosAttributes', kerberosAttributes_aDDomainJoinPassword - The Active Directory password for ADDomainJoinUser.

$sel:crossRealmTrustPrincipalPassword:KerberosAttributes', kerberosAttributes_crossRealmTrustPrincipalPassword - Required only when establishing a cross-realm trust with a KDC in a different realm. The cross-realm principal password, which must be identical across realms.

$sel:aDDomainJoinUser:KerberosAttributes', kerberosAttributes_aDDomainJoinUser - Required only when establishing a cross-realm trust with an Active Directory domain. A user with sufficient privileges to join resources to the domain.

kerberosAttributes_kdcAdminPassword :: Lens' KerberosAttributes (Maybe Text)

The password used within the cluster for the kadmin service on the cluster-dedicated KDC, which maintains Kerberos principals, password policies, and keytabs for the cluster.

kerberosAttributes_realm :: Lens' KerberosAttributes (Maybe Text)

The name of the Kerberos realm to which all nodes in a cluster belong. For example, EC2.INTERNAL.

kerberosAttributes_aDDomainJoinPassword :: Lens' KerberosAttributes (Maybe Text)

The Active Directory password for ADDomainJoinUser.

kerberosAttributes_crossRealmTrustPrincipalPassword :: Lens' KerberosAttributes (Maybe Text)

Required only when establishing a cross-realm trust with a KDC in a different realm. The cross-realm principal password, which must be identical across realms.

kerberosAttributes_aDDomainJoinUser :: Lens' KerberosAttributes (Maybe Text)

Required only when establishing a cross-realm trust with an Active Directory domain. A user with sufficient privileges to join resources to the domain.