Network.AWS.EMR.Types.ManagedScalingPolicy

Description

Documentation

data ManagedScalingPolicy

Managed scaling policy for an Amazon EMR cluster. The policy specifies the limits for resources that can be added or terminated from a cluster. The policy only applies to the core and task nodes. The master node cannot be scaled after initial configuration.

See: newManagedScalingPolicy smart constructor.

Constructors

ManagedScalingPolicy'

Fields

computeLimits :: Maybe ComputeLimits

The EC2 unit limits for a managed scaling policy. The managed scaling activity of a cluster is not allowed to go above or below these limits. The limit only applies to the core and task nodes. The master node cannot be scaled after initial configuration.

newManagedScalingPolicy :: ManagedScalingPolicy

Create a value of ManagedScalingPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:computeLimits:ManagedScalingPolicy', managedScalingPolicy_computeLimits - The EC2 unit limits for a managed scaling policy. The managed scaling activity of a cluster is not allowed to go above or below these limits. The limit only applies to the core and task nodes. The master node cannot be scaled after initial configuration.

managedScalingPolicy_computeLimits :: Lens' ManagedScalingPolicy (Maybe ComputeLimits)

The EC2 unit limits for a managed scaling policy. The managed scaling activity of a cluster is not allowed to go above or below these limits. The limit only applies to the core and task nodes. The master node cannot be scaled after initial configuration.