Network.AWS.EMR.Types.NotebookExecution

Description

Documentation

data NotebookExecution

A notebook execution. An execution is a specific instance that an EMR Notebook is run using the StartNotebookExecution action.

See: newNotebookExecution smart constructor.

Constructors

NotebookExecution'

Fields

lastStateChangeReason :: Maybe Text

The reason for the latest status change of the notebook execution.

status :: Maybe NotebookExecutionStatus

The status of the notebook execution.

START_PENDING indicates that the cluster has received the execution request but execution has not begun.

STARTING indicates that the execution is starting on the cluster.

RUNNING indicates that the execution is being processed by the cluster.

FINISHING indicates that execution processing is in the final stages.

FINISHED indicates that the execution has completed without error.

FAILING indicates that the execution is failing and will not finish successfully.

FAILED indicates that the execution failed.

STOP_PENDING indicates that the cluster has received a StopNotebookExecution request and the stop is pending.

STOPPING indicates that the cluster is in the process of stopping the execution as a result of a StopNotebookExecution request.

STOPPED indicates that the execution stopped because of a StopNotebookExecution request.

notebookExecutionName :: Maybe Text

A name for the notebook execution.

editorId :: Maybe Text

The unique identifier of the EMR Notebook that is used for the notebook execution.

notebookExecutionId :: Maybe Text

The unique identifier of a notebook execution.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the notebook execution.

startTime :: Maybe POSIX

The timestamp when notebook execution started.

notebookParams :: Maybe Text

Input parameters in JSON format passed to the EMR Notebook at runtime for execution.

endTime :: Maybe POSIX

The timestamp when notebook execution ended.

executionEngine :: Maybe ExecutionEngineConfig

The execution engine, such as an EMR cluster, used to run the EMR notebook and perform the notebook execution.

notebookInstanceSecurityGroupId :: Maybe Text

The unique identifier of the EC2 security group associated with the EMR Notebook instance. For more information see Specifying EC2 Security Groups for EMR Notebooks in the EMR Management Guide.

tags :: Maybe [Tag]

A list of tags associated with a notebook execution. Tags are user-defined key-value pairs that consist of a required key string with a maximum of 128 characters and an optional value string with a maximum of 256 characters.

outputNotebookURI :: Maybe Text

The location of the notebook execution's output file in Amazon S3.

newNotebookExecution :: NotebookExecution

Create a value of NotebookExecution with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastStateChangeReason:NotebookExecution', notebookExecution_lastStateChangeReason - The reason for the latest status change of the notebook execution.

$sel:status:NotebookExecution', notebookExecution_status - The status of the notebook execution.

START_PENDING indicates that the cluster has received the execution request but execution has not begun.

STARTING indicates that the execution is starting on the cluster.

RUNNING indicates that the execution is being processed by the cluster.

FINISHING indicates that execution processing is in the final stages.

FINISHED indicates that the execution has completed without error.

FAILING indicates that the execution is failing and will not finish successfully.

FAILED indicates that the execution failed.

STOP_PENDING indicates that the cluster has received a StopNotebookExecution request and the stop is pending.

STOPPING indicates that the cluster is in the process of stopping the execution as a result of a StopNotebookExecution request.

STOPPED indicates that the execution stopped because of a StopNotebookExecution request.

$sel:notebookExecutionName:NotebookExecution', notebookExecution_notebookExecutionName - A name for the notebook execution.

$sel:editorId:NotebookExecution', notebookExecution_editorId - The unique identifier of the EMR Notebook that is used for the notebook execution.

$sel:notebookExecutionId:NotebookExecution', notebookExecution_notebookExecutionId - The unique identifier of a notebook execution.

$sel:arn:NotebookExecution', notebookExecution_arn - The Amazon Resource Name (ARN) of the notebook execution.

$sel:startTime:NotebookExecution', notebookExecution_startTime - The timestamp when notebook execution started.

$sel:notebookParams:NotebookExecution', notebookExecution_notebookParams - Input parameters in JSON format passed to the EMR Notebook at runtime for execution.

$sel:endTime:NotebookExecution', notebookExecution_endTime - The timestamp when notebook execution ended.

$sel:executionEngine:NotebookExecution', notebookExecution_executionEngine - The execution engine, such as an EMR cluster, used to run the EMR notebook and perform the notebook execution.

$sel:notebookInstanceSecurityGroupId:NotebookExecution', notebookExecution_notebookInstanceSecurityGroupId - The unique identifier of the EC2 security group associated with the EMR Notebook instance. For more information see Specifying EC2 Security Groups for EMR Notebooks in the EMR Management Guide.

$sel:tags:NotebookExecution', notebookExecution_tags - A list of tags associated with a notebook execution. Tags are user-defined key-value pairs that consist of a required key string with a maximum of 128 characters and an optional value string with a maximum of 256 characters.

$sel:outputNotebookURI:NotebookExecution', notebookExecution_outputNotebookURI - The location of the notebook execution's output file in Amazon S3.

notebookExecution_lastStateChangeReason :: Lens' NotebookExecution (Maybe Text)

The reason for the latest status change of the notebook execution.

notebookExecution_status :: Lens' NotebookExecution (Maybe NotebookExecutionStatus)

The status of the notebook execution.

START_PENDING indicates that the cluster has received the execution request but execution has not begun.

STARTING indicates that the execution is starting on the cluster.

RUNNING indicates that the execution is being processed by the cluster.

FINISHING indicates that execution processing is in the final stages.

FINISHED indicates that the execution has completed without error.

FAILING indicates that the execution is failing and will not finish successfully.

FAILED indicates that the execution failed.

STOP_PENDING indicates that the cluster has received a StopNotebookExecution request and the stop is pending.

STOPPING indicates that the cluster is in the process of stopping the execution as a result of a StopNotebookExecution request.

STOPPED indicates that the execution stopped because of a StopNotebookExecution request.

notebookExecution_notebookExecutionName :: Lens' NotebookExecution (Maybe Text)

A name for the notebook execution.

notebookExecution_editorId :: Lens' NotebookExecution (Maybe Text)

The unique identifier of the EMR Notebook that is used for the notebook execution.

notebookExecution_notebookExecutionId :: Lens' NotebookExecution (Maybe Text)

The unique identifier of a notebook execution.

notebookExecution_arn :: Lens' NotebookExecution (Maybe Text)

The Amazon Resource Name (ARN) of the notebook execution.

notebookExecution_startTime :: Lens' NotebookExecution (Maybe UTCTime)

The timestamp when notebook execution started.

notebookExecution_notebookParams :: Lens' NotebookExecution (Maybe Text)

Input parameters in JSON format passed to the EMR Notebook at runtime for execution.

notebookExecution_endTime :: Lens' NotebookExecution (Maybe UTCTime)

The timestamp when notebook execution ended.

notebookExecution_executionEngine :: Lens' NotebookExecution (Maybe ExecutionEngineConfig)

The execution engine, such as an EMR cluster, used to run the EMR notebook and perform the notebook execution.

notebookExecution_notebookInstanceSecurityGroupId :: Lens' NotebookExecution (Maybe Text)

The unique identifier of the EC2 security group associated with the EMR Notebook instance. For more information see Specifying EC2 Security Groups for EMR Notebooks in the EMR Management Guide.

notebookExecution_tags :: Lens' NotebookExecution (Maybe [Tag])

A list of tags associated with a notebook execution. Tags are user-defined key-value pairs that consist of a required key string with a maximum of 128 characters and an optional value string with a maximum of 256 characters.

notebookExecution_outputNotebookURI :: Lens' NotebookExecution (Maybe Text)

The location of the notebook execution's output file in Amazon S3.