Network.AWS.EMR.Types.NotebookExecutionSummary

Description

Documentation

data NotebookExecutionSummary

Details for a notebook execution. The details include information such as the unique ID and status of the notebook execution.

See: newNotebookExecutionSummary smart constructor.

Constructors

NotebookExecutionSummary'

Fields

status :: Maybe NotebookExecutionStatus

The status of the notebook execution.

START_PENDING indicates that the cluster has received the execution request but execution has not begun.

STARTING indicates that the execution is starting on the cluster.

RUNNING indicates that the execution is being processed by the cluster.

FINISHING indicates that execution processing is in the final stages.

FINISHED indicates that the execution has completed without error.

FAILING indicates that the execution is failing and will not finish successfully.

FAILED indicates that the execution failed.

STOP_PENDING indicates that the cluster has received a StopNotebookExecution request and the stop is pending.

STOPPING indicates that the cluster is in the process of stopping the execution as a result of a StopNotebookExecution request.

STOPPED indicates that the execution stopped because of a StopNotebookExecution request.

notebookExecutionName :: Maybe Text

The name of the notebook execution.

editorId :: Maybe Text

The unique identifier of the editor associated with the notebook execution.

notebookExecutionId :: Maybe Text

The unique identifier of the notebook execution.

startTime :: Maybe POSIX

The timestamp when notebook execution started.

endTime :: Maybe POSIX

The timestamp when notebook execution started.

newNotebookExecutionSummary :: NotebookExecutionSummary

Create a value of NotebookExecutionSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:NotebookExecutionSummary', notebookExecutionSummary_status - The status of the notebook execution.

START_PENDING indicates that the cluster has received the execution request but execution has not begun.

STARTING indicates that the execution is starting on the cluster.

RUNNING indicates that the execution is being processed by the cluster.

FINISHING indicates that execution processing is in the final stages.

FINISHED indicates that the execution has completed without error.

FAILING indicates that the execution is failing and will not finish successfully.

FAILED indicates that the execution failed.

STOP_PENDING indicates that the cluster has received a StopNotebookExecution request and the stop is pending.

STOPPING indicates that the cluster is in the process of stopping the execution as a result of a StopNotebookExecution request.

STOPPED indicates that the execution stopped because of a StopNotebookExecution request.

$sel:notebookExecutionName:NotebookExecutionSummary', notebookExecutionSummary_notebookExecutionName - The name of the notebook execution.

$sel:editorId:NotebookExecutionSummary', notebookExecutionSummary_editorId - The unique identifier of the editor associated with the notebook execution.

$sel:notebookExecutionId:NotebookExecutionSummary', notebookExecutionSummary_notebookExecutionId - The unique identifier of the notebook execution.

$sel:startTime:NotebookExecutionSummary', notebookExecutionSummary_startTime - The timestamp when notebook execution started.

$sel:endTime:NotebookExecutionSummary', notebookExecutionSummary_endTime - The timestamp when notebook execution started.

notebookExecutionSummary_status :: Lens' NotebookExecutionSummary (Maybe NotebookExecutionStatus)

The status of the notebook execution.

START_PENDING indicates that the cluster has received the execution request but execution has not begun.

STARTING indicates that the execution is starting on the cluster.

RUNNING indicates that the execution is being processed by the cluster.

FINISHING indicates that execution processing is in the final stages.

FINISHED indicates that the execution has completed without error.

FAILING indicates that the execution is failing and will not finish successfully.

FAILED indicates that the execution failed.

STOP_PENDING indicates that the cluster has received a StopNotebookExecution request and the stop is pending.

STOPPING indicates that the cluster is in the process of stopping the execution as a result of a StopNotebookExecution request.

STOPPED indicates that the execution stopped because of a StopNotebookExecution request.

notebookExecutionSummary_notebookExecutionName :: Lens' NotebookExecutionSummary (Maybe Text)

The name of the notebook execution.

notebookExecutionSummary_editorId :: Lens' NotebookExecutionSummary (Maybe Text)

The unique identifier of the editor associated with the notebook execution.

notebookExecutionSummary_notebookExecutionId :: Lens' NotebookExecutionSummary (Maybe Text)

The unique identifier of the notebook execution.

notebookExecutionSummary_startTime :: Lens' NotebookExecutionSummary (Maybe UTCTime)

The timestamp when notebook execution started.

notebookExecutionSummary_endTime :: Lens' NotebookExecutionSummary (Maybe UTCTime)

The timestamp when notebook execution started.