Network.AWS.EMR.Types.OnDemandProvisioningSpecification

Description

Documentation

data OnDemandProvisioningSpecification

The launch specification for On-Demand Instances in the instance fleet, which determines the allocation strategy.

The instance fleet configuration is available only in Amazon EMR versions 4.8.0 and later, excluding 5.0.x versions. On-Demand Instances allocation strategy is available in Amazon EMR version 5.12.1 and later.

See: newOnDemandProvisioningSpecification smart constructor.

Constructors

OnDemandProvisioningSpecification'

Fields

capacityReservationOptions :: Maybe OnDemandCapacityReservationOptions

The launch specification for On-Demand instances in the instance fleet, which determines the allocation strategy.

allocationStrategy :: OnDemandProvisioningAllocationStrategy

Specifies the strategy to use in launching On-Demand instance fleets. Currently, the only option is lowest-price (the default), which launches the lowest price first.

newOnDemandProvisioningSpecification

Arguments

:: OnDemandProvisioningAllocationStrategy
$sel:allocationStrategy:OnDemandProvisioningSpecification'

-> OnDemandProvisioningSpecification

Create a value of OnDemandProvisioningSpecification with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:capacityReservationOptions:OnDemandProvisioningSpecification', onDemandProvisioningSpecification_capacityReservationOptions - The launch specification for On-Demand instances in the instance fleet, which determines the allocation strategy.

$sel:allocationStrategy:OnDemandProvisioningSpecification', onDemandProvisioningSpecification_allocationStrategy - Specifies the strategy to use in launching On-Demand instance fleets. Currently, the only option is lowest-price (the default), which launches the lowest price first.

onDemandProvisioningSpecification_capacityReservationOptions :: Lens' OnDemandProvisioningSpecification (Maybe OnDemandCapacityReservationOptions)

The launch specification for On-Demand instances in the instance fleet, which determines the allocation strategy.

onDemandProvisioningSpecification_allocationStrategy :: Lens' OnDemandProvisioningSpecification OnDemandProvisioningAllocationStrategy

Specifies the strategy to use in launching On-Demand instance fleets. Currently, the only option is lowest-price (the default), which launches the lowest price first.