Network.AWS.EMR.Types.Step

Description

Documentation

data Step

This represents a step in a cluster.

See: newStep smart constructor.

Constructors

Step'

Fields

status :: Maybe StepStatus

The current execution status details of the cluster step.

id :: Maybe Text

The identifier of the cluster step.

config :: Maybe HadoopStepConfig

The Hadoop job configuration of the cluster step.

name :: Maybe Text

The name of the cluster step.

actionOnFailure :: Maybe ActionOnFailure

The action to take when the cluster step fails. Possible values are TERMINATE_CLUSTER, CANCEL_AND_WAIT, and CONTINUE. TERMINATE_JOB_FLOW is provided for backward compatibility. We recommend using TERMINATE_CLUSTER instead.

If a cluster's StepConcurrencyLevel is greater than 1, do not use AddJobFlowSteps to submit a step with this parameter set to CANCEL_AND_WAIT or TERMINATE_CLUSTER. The step is not submitted and the action fails with a message that the ActionOnFailure setting is not valid.

If you change a cluster's StepConcurrencyLevel to be greater than 1 while a step is running, the ActionOnFailure parameter may not behave as you expect. In this case, for a step that fails with this parameter set to CANCEL_AND_WAIT, pending steps and the running step are not canceled; for a step that fails with this parameter set to TERMINATE_CLUSTER, the cluster does not terminate.

newStep :: Step

Create a value of Step with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Step', step_status - The current execution status details of the cluster step.

$sel:id:Step', step_id - The identifier of the cluster step.

$sel:config:Step', step_config - The Hadoop job configuration of the cluster step.

$sel:name:Step', step_name - The name of the cluster step.

$sel:actionOnFailure:Step', step_actionOnFailure - The action to take when the cluster step fails. Possible values are TERMINATE_CLUSTER, CANCEL_AND_WAIT, and CONTINUE. TERMINATE_JOB_FLOW is provided for backward compatibility. We recommend using TERMINATE_CLUSTER instead.

If a cluster's StepConcurrencyLevel is greater than 1, do not use AddJobFlowSteps to submit a step with this parameter set to CANCEL_AND_WAIT or TERMINATE_CLUSTER. The step is not submitted and the action fails with a message that the ActionOnFailure setting is not valid.

If you change a cluster's StepConcurrencyLevel to be greater than 1 while a step is running, the ActionOnFailure parameter may not behave as you expect. In this case, for a step that fails with this parameter set to CANCEL_AND_WAIT, pending steps and the running step are not canceled; for a step that fails with this parameter set to TERMINATE_CLUSTER, the cluster does not terminate.

step_status :: Lens' Step (Maybe StepStatus)

The current execution status details of the cluster step.

step_id :: Lens' Step (Maybe Text)

The identifier of the cluster step.

step_config :: Lens' Step (Maybe HadoopStepConfig)

The Hadoop job configuration of the cluster step.

step_name :: Lens' Step (Maybe Text)

The name of the cluster step.

step_actionOnFailure :: Lens' Step (Maybe ActionOnFailure)

The action to take when the cluster step fails. Possible values are TERMINATE_CLUSTER, CANCEL_AND_WAIT, and CONTINUE. TERMINATE_JOB_FLOW is provided for backward compatibility. We recommend using TERMINATE_CLUSTER instead.

If a cluster's StepConcurrencyLevel is greater than 1, do not use AddJobFlowSteps to submit a step with this parameter set to CANCEL_AND_WAIT or TERMINATE_CLUSTER. The step is not submitted and the action fails with a message that the ActionOnFailure setting is not valid.

If you change a cluster's StepConcurrencyLevel to be greater than 1 while a step is running, the ActionOnFailure parameter may not behave as you expect. In this case, for a step that fails with this parameter set to CANCEL_AND_WAIT, pending steps and the running step are not canceled; for a step that fails with this parameter set to TERMINATE_CLUSTER, the cluster does not terminate.