Network.AWS.EMR.Types.StepStateChangeReason

Description

Documentation

data StepStateChangeReason

The details of the step state change reason.

See: newStepStateChangeReason smart constructor.

Constructors

StepStateChangeReason'

Fields

code :: Maybe StepStateChangeReasonCode

The programmable code for the state change reason. Note: Currently, the service provides no code for the state change.

message :: Maybe Text

The descriptive message for the state change reason.

newStepStateChangeReason :: StepStateChangeReason

Create a value of StepStateChangeReason with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:code:StepStateChangeReason', stepStateChangeReason_code - The programmable code for the state change reason. Note: Currently, the service provides no code for the state change.

$sel:message:StepStateChangeReason', stepStateChangeReason_message - The descriptive message for the state change reason.

stepStateChangeReason_code :: Lens' StepStateChangeReason (Maybe StepStateChangeReasonCode)

The programmable code for the state change reason. Note: Currently, the service provides no code for the state change.

stepStateChangeReason_message :: Lens' StepStateChangeReason (Maybe Text)

The descriptive message for the state change reason.