Network.AWS.EMR.Types.StepStatus

Description

Documentation

data StepStatus

The execution status details of the cluster step.

See: newStepStatus smart constructor.

Constructors

StepStatus'

Fields

stateChangeReason :: Maybe StepStateChangeReason

The reason for the step execution status change.

failureDetails :: Maybe FailureDetails

The details for the step failure including reason, message, and log file path where the root cause was identified.

state :: Maybe StepState

The execution state of the cluster step.

timeline :: Maybe StepTimeline

The timeline of the cluster step status over time.

newStepStatus :: StepStatus

Create a value of StepStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stateChangeReason:StepStatus', stepStatus_stateChangeReason - The reason for the step execution status change.

$sel:failureDetails:StepStatus', stepStatus_failureDetails - The details for the step failure including reason, message, and log file path where the root cause was identified.

$sel:state:StepStatus', stepStatus_state - The execution state of the cluster step.

$sel:timeline:StepStatus', stepStatus_timeline - The timeline of the cluster step status over time.

stepStatus_stateChangeReason :: Lens' StepStatus (Maybe StepStateChangeReason)

The reason for the step execution status change.

stepStatus_failureDetails :: Lens' StepStatus (Maybe FailureDetails)

The details for the step failure including reason, message, and log file path where the root cause was identified.

stepStatus_state :: Lens' StepStatus (Maybe StepState)

The execution state of the cluster step.

stepStatus_timeline :: Lens' StepStatus (Maybe StepTimeline)

The timeline of the cluster step status over time.