Network.AWS.EMR.Types.StudioSummary

Description

Documentation

data StudioSummary

Details for an Amazon EMR Studio, including ID, Name, VPC, and Description. The details do not include subnets, IAM roles, security groups, or tags associated with the Studio.

See: newStudioSummary smart constructor.

Constructors

StudioSummary'

Fields

creationTime :: Maybe POSIX

The time when the Amazon EMR Studio was created.

studioId :: Maybe Text

The ID of the Amazon EMR Studio.

vpcId :: Maybe Text

The ID of the Virtual Private Cloud (Amazon VPC) associated with the Amazon EMR Studio.

url :: Maybe Text

The unique access URL of the Amazon EMR Studio.

authMode :: Maybe AuthMode

Specifies whether the Studio authenticates users using IAM or Amazon Web Services SSO.

name :: Maybe Text

The name of the Amazon EMR Studio.

description :: Maybe Text

The detailed description of the Amazon EMR Studio.

newStudioSummary :: StudioSummary

Create a value of StudioSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:StudioSummary', studioSummary_creationTime - The time when the Amazon EMR Studio was created.

$sel:studioId:StudioSummary', studioSummary_studioId - The ID of the Amazon EMR Studio.

$sel:vpcId:StudioSummary', studioSummary_vpcId - The ID of the Virtual Private Cloud (Amazon VPC) associated with the Amazon EMR Studio.

$sel:url:StudioSummary', studioSummary_url - The unique access URL of the Amazon EMR Studio.

$sel:authMode:StudioSummary', studioSummary_authMode - Specifies whether the Studio authenticates users using IAM or Amazon Web Services SSO.

$sel:name:StudioSummary', studioSummary_name - The name of the Amazon EMR Studio.

$sel:description:StudioSummary', studioSummary_description - The detailed description of the Amazon EMR Studio.

studioSummary_creationTime :: Lens' StudioSummary (Maybe UTCTime)

The time when the Amazon EMR Studio was created.

studioSummary_studioId :: Lens' StudioSummary (Maybe Text)

The ID of the Amazon EMR Studio.

studioSummary_vpcId :: Lens' StudioSummary (Maybe Text)

The ID of the Virtual Private Cloud (Amazon VPC) associated with the Amazon EMR Studio.

studioSummary_url :: Lens' StudioSummary (Maybe Text)

The unique access URL of the Amazon EMR Studio.

studioSummary_authMode :: Lens' StudioSummary (Maybe AuthMode)

Specifies whether the Studio authenticates users using IAM or Amazon Web Services SSO.

studioSummary_name :: Lens' StudioSummary (Maybe Text)

The name of the Amazon EMR Studio.

studioSummary_description :: Lens' StudioSummary (Maybe Text)

The detailed description of the Amazon EMR Studio.