Network.AWS.EMR.UpdateStudio

Description

Updates an Amazon EMR Studio configuration, including attributes such as name, description, and subnets.

Creating a Request

data UpdateStudio

See: newUpdateStudio smart constructor.

Constructors

UpdateStudio'

Fields

defaultS3Location :: Maybe Text

The Amazon S3 location to back up Workspaces and notebook files for the Amazon EMR Studio.

subnetIds :: Maybe [Text]

A list of subnet IDs to associate with the Amazon EMR Studio. The list can include new subnet IDs, but must also include all of the subnet IDs previously associated with the Studio. The list order does not matter. A Studio can have a maximum of 5 subnets. The subnets must belong to the same VPC as the Studio.

name :: Maybe Text

A descriptive name for the Amazon EMR Studio.

description :: Maybe Text

A detailed description to assign to the Amazon EMR Studio.

studioId :: Text

The ID of the Amazon EMR Studio to update.

newUpdateStudio

Arguments

:: Text
$sel:studioId:UpdateStudio'

-> UpdateStudio

Create a value of UpdateStudio with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:defaultS3Location:UpdateStudio', updateStudio_defaultS3Location - The Amazon S3 location to back up Workspaces and notebook files for the Amazon EMR Studio.

$sel:subnetIds:UpdateStudio', updateStudio_subnetIds - A list of subnet IDs to associate with the Amazon EMR Studio. The list can include new subnet IDs, but must also include all of the subnet IDs previously associated with the Studio. The list order does not matter. A Studio can have a maximum of 5 subnets. The subnets must belong to the same VPC as the Studio.

$sel:name:UpdateStudio', updateStudio_name - A descriptive name for the Amazon EMR Studio.

$sel:description:UpdateStudio', updateStudio_description - A detailed description to assign to the Amazon EMR Studio.

$sel:studioId:UpdateStudio', updateStudio_studioId - The ID of the Amazon EMR Studio to update.

Request Lenses

updateStudio_defaultS3Location :: Lens' UpdateStudio (Maybe Text)

The Amazon S3 location to back up Workspaces and notebook files for the Amazon EMR Studio.

updateStudio_subnetIds :: Lens' UpdateStudio (Maybe [Text])

A list of subnet IDs to associate with the Amazon EMR Studio. The list can include new subnet IDs, but must also include all of the subnet IDs previously associated with the Studio. The list order does not matter. A Studio can have a maximum of 5 subnets. The subnets must belong to the same VPC as the Studio.

updateStudio_name :: Lens' UpdateStudio (Maybe Text)

A descriptive name for the Amazon EMR Studio.

updateStudio_description :: Lens' UpdateStudio (Maybe Text)

A detailed description to assign to the Amazon EMR Studio.

updateStudio_studioId :: Lens' UpdateStudio Text

The ID of the Amazon EMR Studio to update.

Destructuring the Response

data UpdateStudioResponse

See: newUpdateStudioResponse smart constructor.

Constructors

UpdateStudioResponse'

newUpdateStudioResponse :: UpdateStudioResponse

Create a value of UpdateStudioResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.