Network.AWS.EMR.V2009_03_31

Description

Amazon Elastic MapReduce (Amazon EMR) is a web service that makes it easy to process large amounts of data efficiently. Amazon EMR uses Hadoop processing combined with several AWS products to do such tasks as web indexing, data mining, log file analysis, machine learning, scientific simulation, and data warehousing.

Documentation

data AddInstanceGroups

Constructors

AddInstanceGroups

Fields

_aigiInstanceGroups :: [InstanceGroupConfig]

Instance Groups to add.

_aigiJobFlowId :: Text

Job flow in which to add the instance groups.

aigiJobFlowId :: Lens' AddInstanceGroups Text

aigiInstanceGroups :: Lens' AddInstanceGroups [InstanceGroupConfig]

data AddInstanceGroupsResponse

Constructors

AddInstanceGroupsResponse

Fields

_aigoInstanceGroupIds :: [Text]

Instance group IDs of the newly created instance groups.

_aigoJobFlowId :: Maybe Text

The job flow ID in which the instance groups are added.

aigoJobFlowId :: Lens' AddInstanceGroupsResponse (Maybe Text)

aigoInstanceGroupIds :: Lens' AddInstanceGroupsResponse [Text]

data AddJobFlowSteps

Constructors

AddJobFlowSteps

Fields

_ajfsiSteps :: [StepConfig]

A list of StepConfig to be executed by the job flow.

_ajfsiJobFlowId :: Text

A string that uniquely identifies the job flow. This identifier is returned by RunJobFlow and can also be obtained from ListClusters.

ajfsiSteps :: Lens' AddJobFlowSteps [StepConfig]

ajfsiJobFlowId :: Lens' AddJobFlowSteps Text

data AddJobFlowStepsResponse

Constructors

AddJobFlowStepsResponse

Fields

_ajfsoStepIds :: [Text]

The identifiers of the list of steps added to the job flow.

ajfsoStepIds :: Iso' AddJobFlowStepsResponse [Text]

data AddTags

Constructors

AddTags

Fields

_atiResourceId :: Text

The Amazon EMR resource identifier to which tags will be added. This value must be a cluster identifier.

_atiTags :: [Tag]

A list of tags to associate with a cluster and propagate to Amazon EC2 instances. Tags are user-defined key/value pairs that consist of a required key string with a maximum of 128 characters, and an optional value string with a maximum of 256 characters.

atiTags :: Lens' AddTags [Tag]

atiResourceId :: Lens' AddTags Text

data AddTagsResponse

Constructors

AddTagsResponse

data DescribeCluster

Constructors

DescribeCluster

Fields

_dciClusterId :: Text

The identifier of the cluster to describe.

dciClusterId :: Iso' DescribeCluster Text

data DescribeClusterResponse

Constructors

DescribeClusterResponse

Fields

_dcoCluster :: Maybe Cluster

This output contains the details for the requested cluster.

dcoCluster :: Iso' DescribeClusterResponse (Maybe Cluster)

describeJobFlows :: DescribeJobFlows

Minimum specification for a DescribeJobFlows request.

data DescribeJobFlows

Constructors

DescribeJobFlows

Fields

_djfiCreatedAfter :: Maybe POSIX

Return only job flows created after this date and time.

_djfiCreatedBefore :: Maybe POSIX

Return only job flows created before this date and time.

_djfiJobFlowStates :: [JobFlowExecutionState]

Return only job flows whose state is contained in this list.

_djfiJobFlowIds :: [Text]

Return only job flows whose job flow ID is contained in this list.

djfiJobFlowStates :: Lens' DescribeJobFlows [JobFlowExecutionState]

djfiJobFlowIds :: Lens' DescribeJobFlows [Text]

djfiCreatedBefore :: Lens' DescribeJobFlows (Maybe POSIX)

djfiCreatedAfter :: Lens' DescribeJobFlows (Maybe POSIX)

data DescribeJobFlowsResponse

Constructors

DescribeJobFlowsResponse

Fields

_djfoJobFlows :: [JobFlowDetail]

A list of job flows matching the parameters supplied.

djfoJobFlows :: Iso' DescribeJobFlowsResponse [JobFlowDetail]

data DescribeStep

Constructors

DescribeStep

Fields

_dsiClusterId :: Text

The identifier of the cluster with steps to describe.

_dsiStepId :: Text

The identifier of the step to describe.

dsiStepId :: Lens' DescribeStep Text

dsiClusterId :: Lens' DescribeStep Text

data DescribeStepResponse

Constructors

DescribeStepResponse

Fields

_dsoStep :: Maybe Step

The step details for the requested step identifier.

dsoStep :: Iso' DescribeStepResponse (Maybe Step)

listBootstrapActions

Arguments

:: Text
_lbaiClusterId

-> ListBootstrapActions

Minimum specification for a ListBootstrapActions request.

data ListBootstrapActions

Constructors

ListBootstrapActions

Fields

_lbaiClusterId :: Text

The cluster identifier for the bootstrap actions to list .

_lbaiMarker :: Maybe Text

The pagination token that indicates the next set of results to retrieve .

lbaiMarker :: Lens' ListBootstrapActions (Maybe Text)

lbaiClusterId :: Lens' ListBootstrapActions Text

data ListBootstrapActionsResponse

Constructors

ListBootstrapActionsResponse

Fields

_lbaoBootstrapActions :: [Command]

The bootstrap actions associated with the cluster .

_lbaoMarker :: Maybe Text

The pagination token that indicates the next set of results to retrieve .

lbaoMarker :: Lens' ListBootstrapActionsResponse (Maybe Text)

lbaoBootstrapActions :: Lens' ListBootstrapActionsResponse [Command]

listClusters :: ListClusters

Minimum specification for a ListClusters request.

data ListClusters

Constructors

ListClusters

Fields

_lciClusterStates :: [ClusterState]

The cluster state filters to apply when listing clusters.

_lciCreatedAfter :: Maybe POSIX

The creation date and time beginning value filter for listing clusters .

_lciCreatedBefore :: Maybe POSIX

The creation date and time end value filter for listing clusters .

_lciMarker :: Maybe Text

The pagination token that indicates the next set of results to retrieve.

lciMarker :: Lens' ListClusters (Maybe Text)

lciCreatedBefore :: Lens' ListClusters (Maybe POSIX)

lciCreatedAfter :: Lens' ListClusters (Maybe POSIX)

lciClusterStates :: Lens' ListClusters [ClusterState]

data ListClustersResponse

Constructors

ListClustersResponse

Fields

_lcoClusters :: [ClusterSummary]

The list of clusters for the account based on the given filters.

_lcoMarker :: Maybe Text

The pagination token that indicates the next set of results to retrieve.

lcoMarker :: Lens' ListClustersResponse (Maybe Text)

lcoClusters :: Lens' ListClustersResponse [ClusterSummary]

listInstanceGroups

Arguments

:: Text
_ligiClusterId

-> ListInstanceGroups

Minimum specification for a ListInstanceGroups request.

data ListInstanceGroups

Constructors

ListInstanceGroups

Fields

_ligiClusterId :: Text

The identifier of the cluster for which to list the instance groups.

_ligiMarker :: Maybe Text

The pagination token that indicates the next set of results to retrieve.

ligiMarker :: Lens' ListInstanceGroups (Maybe Text)

ligiClusterId :: Lens' ListInstanceGroups Text

data ListInstanceGroupsResponse

Constructors

ListInstanceGroupsResponse

Fields

_ligoInstanceGroups :: [InstanceGroup]

The list of instance groups for the cluster and given filters.

_ligoMarker :: Maybe Text

The pagination token that indicates the next set of results to retrieve.

ligoMarker :: Lens' ListInstanceGroupsResponse (Maybe Text)

ligoInstanceGroups :: Lens' ListInstanceGroupsResponse [InstanceGroup]

listInstances

Arguments

:: Text
_liiClusterId

-> ListInstances

Minimum specification for a ListInstances request.

data ListInstances

Constructors

ListInstances

Fields

_liiClusterId :: Text

The identifier of the cluster for which to list the instances.

_liiInstanceGroupId :: Maybe Text

The identifier of the instance group for which to list the instances.

_liiInstanceGroupTypes :: [InstanceGroupType]

The type of instance group for which to list the instances.

_liiMarker :: Maybe Text

The pagination token that indicates the next set of results to retrieve.

liiMarker :: Lens' ListInstances (Maybe Text)

liiInstanceGroupTypes :: Lens' ListInstances [InstanceGroupType]

liiInstanceGroupId :: Lens' ListInstances (Maybe Text)

liiClusterId :: Lens' ListInstances Text

data ListInstancesResponse

Constructors

ListInstancesResponse

Fields

_lioInstances :: [Instance]

The list of instances for the cluster and given filters.

_lioMarker :: Maybe Text

The pagination token that indicates the next set of results to retrieve.

lioMarker :: Lens' ListInstancesResponse (Maybe Text)

lioInstances :: Lens' ListInstancesResponse [Instance]

listSteps

Arguments

:: Text
_lsiClusterId

-> ListSteps

Minimum specification for a ListSteps request.

data ListSteps

Constructors

ListSteps

Fields

_lsiClusterId :: Text

The identifier of the cluster for which to list the steps.

_lsiMarker :: Maybe Text

The pagination token that indicates the next set of results to retrieve.

_lsiStepStates :: [StepState]

The filter to limit the step list based on certain states.

lsiStepStates :: Lens' ListSteps [StepState]

lsiMarker :: Lens' ListSteps (Maybe Text)

lsiClusterId :: Lens' ListSteps Text

data ListStepsResponse

Constructors

ListStepsResponse

Fields

_lsoMarker :: Maybe Text

The pagination token that indicates the next set of results to retrieve.

_lsoSteps :: [StepSummary]

The filtered list of steps for the cluster.

lsoSteps :: Lens' ListStepsResponse [StepSummary]

lsoMarker :: Lens' ListStepsResponse (Maybe Text)

modifyInstanceGroups :: ModifyInstanceGroups

Minimum specification for a ModifyInstanceGroups request.

data ModifyInstanceGroups

Constructors

ModifyInstanceGroups

Fields

_migiInstanceGroups :: [InstanceGroupModifyConfig]

Instance groups to change.

migiInstanceGroups :: Iso' ModifyInstanceGroups [InstanceGroupModifyConfig]

data ModifyInstanceGroupsResponse

Constructors

ModifyInstanceGroupsResponse

data RemoveTags

Constructors

RemoveTags

Fields

_rtiResourceId :: Text

The Amazon EMR resource identifier from which tags will be removed. This value must be a cluster identifier.

_rtiTagKeys :: [Text]

A list of tag keys to remove from a resource.

rtiTagKeys :: Lens' RemoveTags [Text]

rtiResourceId :: Lens' RemoveTags Text

data RemoveTagsResponse

Constructors

RemoveTagsResponse

runJobFlow

Arguments

:: JobFlowInstancesConfig
_rjfiInstances

-> Text
_rjfiName

-> RunJobFlow

Minimum specification for a RunJobFlow request.

data RunJobFlow

Constructors

RunJobFlow

Fields

_rjfiInstances :: JobFlowInstancesConfig

A specification of the number and type of Amazon EC2 instances on which to run the job flow.

_rjfiName :: Text

The name of the job flow.

_rjfiVisibleToAllUsers :: Maybe Bool

Whether the job flow is visible to all IAM users of the AWS account associated with the job flow. If this value is set to true, all IAM users of that AWS account can view and (if they have the proper policy permissions set) manage the job flow. If it is set to false, only the IAM user that created the job flow can view and manage it.

_rjfiBootstrapActions :: [BootstrapActionConfig]

A list of bootstrap actions that will be run before Hadoop is started on the cluster nodes.

_rjfiNewSupportedProducts :: [SupportedProductConfig]

A list of strings that indicates third-party software to use with the job flow that accepts a user argument list. EMR accepts and forwards the argument list to the corresponding installation script as bootstrap action arguments. For more information, see Launch a Job Flow on the MapR Distribution for Hadoop. Currently supported values are: "mapr-m3" - launch the job flow using MapR M3 Edition. "mapr-m5" - launch the job flow using MapR M5 Edition. "mapr" with the user arguments specifying "--edition,m3" or "--edition,m5" - launch the job flow using MapR M3 or M5 Edition respectively.

_rjfiSteps :: [StepConfig]

A list of steps to be executed by the job flow.

_rjfiSupportedProducts :: [Text]

A list of strings that indicates third-party software to use with the job flow. For more information, go to Use Third Party Applications with Amazon EMR. Currently supported values are: "mapr-m3" - launch the job flow using MapR M3 Edition. "mapr-m5" - launch the job flow using MapR M5 Edition.

_rjfiTags :: [Tag]

A list of tags to associate with a cluster and propagate to Amazon EC2 instances.

_rjfiAdditionalInfo :: Maybe Text

A JSON string for selecting additional features.

_rjfiJobFlowRole :: Maybe Text

An IAM role for the job flow. The EC2 instances of the job flow assume this role. The default role is EMRJobflowDefault. In order to use the default role, you must have already created it using the CLI.

_rjfiLogUri :: Maybe Text

The location in Amazon S3 to write the log files of the job flow. If a value is not provided, logs are not created.

_rjfiServiceRole :: Maybe Text

IAM role that Amazon ElasticMapReduce will assume to work with AWS resources on your behalf. You may set this parameter to the name of an existing IAM role.

_rjfiAmiVersion :: Maybe Text

The version of the Amazon Machine Image (AMI) to use when launching Amazon EC2 instances in the job flow. The following values are valid: "latest" (uses the latest AMI) The version number of the AMI to use, for example, "2.0" If the AMI supports multiple versions of Hadoop (for example, AMI 1.0 supports both Hadoop 0.18 and 0.20) you can use the JobFlowInstancesConfig HadoopVersion parameter to modify the version of Hadoop from the defaults shown above. For details about the AMI versions currently supported by Amazon Elastic MapReduce, go to AMI Versions Supported in Elastic MapReduce in the Amazon Elastic MapReduce Developer's Guide.

rjfiVisibleToAllUsers :: Lens' RunJobFlow (Maybe Bool)

rjfiTags :: Lens' RunJobFlow [Tag]

rjfiSupportedProducts :: Lens' RunJobFlow [Text]

rjfiSteps :: Lens' RunJobFlow [StepConfig]

rjfiServiceRole :: Lens' RunJobFlow (Maybe Text)

rjfiNewSupportedProducts :: Lens' RunJobFlow [SupportedProductConfig]

rjfiName :: Lens' RunJobFlow Text

rjfiLogUri :: Lens' RunJobFlow (Maybe Text)

rjfiJobFlowRole :: Lens' RunJobFlow (Maybe Text)

rjfiInstances :: Lens' RunJobFlow JobFlowInstancesConfig

rjfiBootstrapActions :: Lens' RunJobFlow [BootstrapActionConfig]

rjfiAmiVersion :: Lens' RunJobFlow (Maybe Text)

rjfiAdditionalInfo :: Lens' RunJobFlow (Maybe Text)

data RunJobFlowResponse

Constructors

RunJobFlowResponse

Fields

_rjfoJobFlowId :: Maybe Text

An unique identifier for the job flow.

rjfoJobFlowId :: Iso' RunJobFlowResponse (Maybe Text)

data SetTerminationProtection

Constructors

SetTerminationProtection

Fields

_stpiTerminationProtected :: Bool

A Boolean that indicates whether to protect the job flow and prevent the Amazon EC2 instances in the cluster from shutting down due to API calls, user intervention, or job-flow error.

_stpiJobFlowIds :: [Text]

A list of strings that uniquely identify the job flows to protect. This identifier is returned by RunJobFlow and can also be obtained from DescribeJobFlows .

stpiTerminationProtected :: Lens' SetTerminationProtection Bool

stpiJobFlowIds :: Lens' SetTerminationProtection [Text]

data SetTerminationProtectionResponse

Constructors

SetTerminationProtectionResponse

data SetVisibleToAllUsers

Constructors

SetVisibleToAllUsers

Fields

_svtauiVisibleToAllUsers :: Bool

Whether the specified job flows are visible to all IAM users of the AWS account associated with the job flow. If this value is set to True, all IAM users of that AWS account can view and, if they have the proper IAM policy permissions set, manage the job flows. If it is set to False, only the IAM user that created a job flow can view and manage it.

_svtauiJobFlowIds :: [Text]

Identifiers of the job flows to receive the new visibility setting.

svtauiVisibleToAllUsers :: Lens' SetVisibleToAllUsers Bool

svtauiJobFlowIds :: Lens' SetVisibleToAllUsers [Text]

data SetVisibleToAllUsersResponse

Constructors

SetVisibleToAllUsersResponse

data TerminateJobFlows

Constructors

TerminateJobFlows

Fields

_tjfiJobFlowIds :: [Text]

A list of job flows to be shutdown.

tjfiJobFlowIds :: Iso' TerminateJobFlows [Text]

data TerminateJobFlowsResponse

Constructors

TerminateJobFlowsResponse

data EMR

Supported version (2009-03-31) of the Amazon Elastic MapReduce service.

data ActionOnFailure

The action to take if the job flow step fails.

Constructors

ActionOnFailureCancelAndWait
CANCEL_AND_WAIT

ActionOnFailureContinue
CONTINUE

ActionOnFailureTerminateCluster
TERMINATE_CLUSTER

ActionOnFailureTerminateJobFlow
TERMINATE_JOB_FLOW

data ClusterState

The current state of the cluster.

Constructors

ClusterStateBootstrapping
BOOTSTRAPPING

ClusterStateRunning
RUNNING

ClusterStateStarting
STARTING

ClusterStateTerminated
TERMINATED

ClusterStateTerminatedWithErrors
TERMINATED_WITH_ERRORS

ClusterStateTerminating
TERMINATING

ClusterStateWaiting
WAITING

data ClusterStateChangeReasonCode

The programmatic code for the state change reason.

Constructors

ClusterStateChangeReasonCodeAllStepsCompleted
ALL_STEPS_COMPLETED

ClusterStateChangeReasonCodeBootstrapFailure
BOOTSTRAP_FAILURE

ClusterStateChangeReasonCodeInstanceFailure
INSTANCE_FAILURE

ClusterStateChangeReasonCodeInternalError
INTERNAL_ERROR

ClusterStateChangeReasonCodeStepFailure
STEP_FAILURE

ClusterStateChangeReasonCodeUserRequest
USER_REQUEST

ClusterStateChangeReasonCodeValidationError
VALIDATION_ERROR

data InstanceGroupState

State of instance group. The following values are deprecated: STARTING, TERMINATED, and FAILED.

Constructors

InstanceGroupStateArrested
ARRESTED

InstanceGroupStateBootstrapping
BOOTSTRAPPING

InstanceGroupStateEnded
ENDED

InstanceGroupStateProvisioning
PROVISIONING

InstanceGroupStateResizing
RESIZING

InstanceGroupStateRunning
RUNNING

InstanceGroupStateShuttingDown
SHUTTING_DOWN

InstanceGroupStateSuspended
SUSPENDED

InstanceGroupStateTerminated
TERMINATED

InstanceGroupStateTerminating
TERMINATING

data InstanceGroupStateChangeReasonCode

The programmable code for the state change reason.

Constructors

InstanceGroupStateChangeReasonCodeClusterTerminated
CLUSTER_TERMINATED

InstanceGroupStateChangeReasonCodeInstanceFailure
INSTANCE_FAILURE

InstanceGroupStateChangeReasonCodeInternalError
INTERNAL_ERROR

InstanceGroupStateChangeReasonCodeValidationError
VALIDATION_ERROR

data InstanceGroupType

The type of the instance group. Valid values are MASTER, CORE or TASK.

Constructors

InstanceGroupTypeCore
CORE

InstanceGroupTypeMaster
MASTER

InstanceGroupTypeTask
TASK

data InstanceRoleType

The role of the instance group in the cluster.

Constructors

InstanceRoleTypeCore
CORE

InstanceRoleTypeMaster
MASTER

InstanceRoleTypeTask
TASK

data InstanceState

The current state of the instance.

Constructors

InstanceStateAwaitingFulfillment
AWAITING_FULFILLMENT

InstanceStateBootstrapping
BOOTSTRAPPING

InstanceStateProvisioning
PROVISIONING

InstanceStateRunning
RUNNING

InstanceStateTerminated
TERMINATED

data InstanceStateChangeReasonCode

The programmable code for the state change reason.

Constructors

InstanceStateChangeReasonCodeBootstrapFailure
BOOTSTRAP_FAILURE

InstanceStateChangeReasonCodeClusterTerminated
CLUSTER_TERMINATED

InstanceStateChangeReasonCodeInstanceFailure
INSTANCE_FAILURE

InstanceStateChangeReasonCodeInternalError
INTERNAL_ERROR

InstanceStateChangeReasonCodeValidationError
VALIDATION_ERROR

data JobFlowExecutionState

The type of instance. A small instance A large instance.

Constructors

JobFlowExecutionStateBootstrapping
BOOTSTRAPPING

JobFlowExecutionStateCompleted
COMPLETED

JobFlowExecutionStateFailed
FAILED

JobFlowExecutionStateRunning
RUNNING

JobFlowExecutionStateShuttingDown
SHUTTING_DOWN

JobFlowExecutionStateStarting
STARTING

JobFlowExecutionStateTerminated
TERMINATED

JobFlowExecutionStateWaiting
WAITING

data MarketType

Market type of the Amazon EC2 instances used to create a cluster node.

Constructors

MarketTypeOnDemand
ON_DEMAND

MarketTypeSpot
SPOT

data StepExecutionState

The state of the job flow step.

Constructors

StepExecutionStateCancelled
CANCELLED

StepExecutionStateCompleted
COMPLETED

StepExecutionStateContinue
CONTINUE

StepExecutionStateFailed
FAILED

StepExecutionStateInterrupted
INTERRUPTED

StepExecutionStatePending
PENDING

StepExecutionStateRunning
RUNNING

data StepState

The execution state of the cluster step.

Constructors

StepStateCancelled
CANCELLED

StepStateCompleted
COMPLETED

StepStateFailed
FAILED

StepStateInterrupted
INTERRUPTED

StepStatePending
PENDING

StepStateRunning
RUNNING

data StepStateChangeReasonCode

The programmable code for the state change reason.

Constructors

StepStateChangeReasonCodeNone
NONE

newtype BootstrapActionDetail

Reports the configuration of a bootstrap action in a job flow.

Constructors

BootstrapActionDetail

Fields

_badBootstrapActionConfig :: Maybe BootstrapActionConfig

A description of the bootstrap action.

newtype PlacementType

The Availability Zone the job flow will run in.

Constructors

PlacementType

Fields

_ptAvailabilityZone :: Text

The Amazon EC2 Availability Zone for the job flow.

data Application

An application is any Amazon or third-party software that you can add to the cluster. This structure contains a list of strings that indicates the software to use with the cluster and accepts a user argument list. Amazon EMR accepts and forwards the argument list to the corresponding installation script as bootstrap action argument. For more information, see Launch a Job Flow on the MapR Distribution for Hadoop. Currently supported values are: "mapr-m3" - launch the job flow using MapR M3 Edition. "mapr-m5" - launch the job flow using MapR M5 Edition. "mapr" with the user arguments specifying "--edition,m3" or "--edition,m5" - launch the job flow using MapR M3 or M5 Edition, respectively.

Constructors

Application

Fields

_rArgs :: [Text]

Arguments for Amazon EMR to pass to the application.

_rAdditionalInfo :: HashMap Text Text

This option is for advanced users only. This is meta information about third-party applications that third-party vendors use for testing purposes.

_rName :: Maybe Text

The name of the application.

_rVersion :: Maybe Text

The version of the application.

data BootstrapActionConfig

Configuration of a bootstrap action.

Constructors

BootstrapActionConfig

Fields

_bacName :: Text

The name of the bootstrap action.

_bacScriptBootstrapAction :: ScriptBootstrapActionConfig

The script run by the bootstrap action.

data Cluster

This output contains the details for the requested cluster.

Constructors

Cluster

Fields

_kStatus :: Maybe ClusterStatus

The current status details about the cluster.

_kRequestedAmiVersion :: Maybe Text

The AMI version requested for this cluster.JobFlowDetail$AmiVersion.-->.

_kEc2InstanceAttributes :: Maybe Ec2InstanceAttributes

Provides information about the EC2 instances in a cluster grouped by category. For example, key name, subnet ID, IAM instance profile, and so on.

_kName :: Maybe Text

The name of the cluster.

_kLogUri :: Maybe Text

The path to the Amazon S3 location where logs for this cluster are stored.

_kId :: Maybe Text

The unique identifier for the cluster.

_kRunningAmiVersion :: Maybe Text

The AMI version running on this cluster. This differs from the requested version only if the requested version is a meta version, such as "latest". JobFlowDetail$AmiVersion.-->.

_kTerminationProtected :: Maybe Bool

Indicates whether Amazon EMR will lock the cluster to prevent the EC2 instances from being terminated by an API call or user intervention, or in the event of a cluster error.

_kVisibleToAllUsers :: Maybe Bool

Indicates whether the job flow is visible to all IAM users of the AWS account associated with the job flow. If this value is set to true, all IAM users of that AWS account can view and manage the job flow if they have the proper policy permissions set. If this value is false, only the IAM user that created the cluster can view and manage it. This value can be changed using the SetVisibleToAllUsers action.

_kAutoTerminate :: Maybe Bool

Specifies whether the cluster should terminate after completing all steps.

_kApplications :: [Application]

The applications installed on this cluster.

_kTags :: [Tag]

A list of tags associated with a cluster.

_kServiceRole :: Maybe Text

The IAM role that was specified when the job flow was launched. Amazon ElasticMapReduce will assume this role to work with AWS resources on your behalf.

data ClusterStateChangeReason

The reason for the cluster status change.

Constructors

ClusterStateChangeReason

Fields

_cscrCode :: Maybe ClusterStateChangeReasonCode

The programmatic code for the state change reason.

_cscrMessage :: Maybe Text

The descriptive message for the state change reason.

data ClusterStatus

The current status details about the cluster.

Constructors

ClusterStatus

Fields

_csState :: Maybe ClusterState

The current state of the cluster.

_csStateChangeReason :: Maybe ClusterStateChangeReason

The reason for the cluster status change.

_csTimeline :: Maybe ClusterTimeline

A timeline that represents the status of a cluster over the lifetime of the cluster.

data ClusterSummary

The summary description of the cluster.

Constructors

ClusterSummary

Fields

_cwStatus :: Maybe ClusterStatus

The details about the current status of the cluster.

_cwName :: Maybe Text

The name of the cluster.

_cwId :: Maybe Text

The unique identifier for the cluster.

data ClusterTimeline

A timeline that represents the status of a cluster over the lifetime of the cluster.

Constructors

ClusterTimeline

Fields

_cuReadyDateTime :: Maybe POSIX

The date and time when the cluster was ready to execute steps.

_cuCreationDateTime :: Maybe POSIX

The creation date and time of the cluster.

_cuEndDateTime :: Maybe POSIX

The date and time when the cluster was terminated.

data Command

An entity describing an executable that runs on a cluster.

Constructors

Command

Fields

_cdArgs :: [Text]

Arguments for Amazon EMR to pass to the command for execution.

_cdScriptPath :: Maybe Text

The Amazon S3 location of the command script.

_cdName :: Maybe Text

The name of the command.

data Ec2InstanceAttributes

Provides information about the EC2 instances in a cluster grouped by category. For example, key name, subnet ID, IAM instance profile, and so on.

Constructors

Ec2InstanceAttributes

Fields

_eiaEc2KeyName :: Maybe Text

The name of the Amazon EC2 key pair to use when connecting with SSH into the master node as a user named "hadoop".

_eiaIamInstanceProfile :: Maybe Text

The IAM role that was specified when the job flow was launched. The EC2 instances of the job flow assume this role.

_eiaEc2SubnetId :: Maybe Text

To launch the job flow in Amazon VPC, set this parameter to the identifier of the Amazon VPC subnet where you want the job flow to launch. If you do not specify this value, the job flow is launched in the normal AWS cloud, outside of a VPC. Amazon VPC currently does not support cluster compute quadruple extra large (cc1.4xlarge) instances. Thus, you cannot specify the cc1.4xlarge instance type for nodes of a job flow launched in a VPC.

_eiaEc2AvailabilityZone :: Maybe Text

The Availability Zone in which the cluster will run.

data HadoopJarStepConfig

The JAR file used for the job flow step.

Constructors

HadoopJarStepConfig

Fields

_hjscArgs :: [Text]

A list of command line arguments passed to the JAR file's main function when executed.

_hjscJar :: Text

A path to a JAR file run during the step.

_hjscMainClass :: Maybe Text

The name of the main class in the specified Java file. If not specified, the JAR file should specify a Main-Class in its manifest file.

_hjscProperties :: [KeyValue]

A list of Java properties that are set when the step runs. You can use these properties to pass key value pairs to your main function.

data HadoopStepConfig

The Hadoop job configuration of the cluster step.

Constructors

HadoopStepConfig

Fields

_hscArgs :: [Text]

The list of command line arguments to pass to the JAR file's main function for execution.

_hscJar :: Maybe Text

The path to the JAR file that runs during the step.

_hscMainClass :: Maybe Text

The name of the main class in the specified Java file. If not specified, the JAR file should specify a main class in its manifest file.

_hscProperties :: HashMap Text Text

The list of Java properties that are set when the step runs. You can use these properties to pass key value pairs to your main function.

data Instance

Represents an EC2 instance provisioned as part of cluster.

Constructors

Instance

Fields

_ieStatus :: Maybe InstanceStatus

The current status of the instance.

_iePublicDnsName :: Maybe Text

The public DNS name of the instance.

_ieEc2InstanceId :: Maybe Text

The unique identifier of the instance in Amazon EC2.

_iePrivateIpAddress :: Maybe Text

The private IP address of the instance.

_ieId :: Maybe Text

The unique identifier for the instance in Amazon EMR.

_iePrivateDnsName :: Maybe Text

The private DNS name of the instance.

_iePublicIpAddress :: Maybe Text

The public IP address of the instance.

data InstanceGroup

This entity represents an instance group, which is a group of instances that have common purpose. For example, CORE instance group is used for HDFS.

Constructors

InstanceGroup

Fields

_igStatus :: Maybe InstanceGroupStatus

The current status of the instance group.

_igBidPrice :: Maybe Text

The bid price for each EC2 instance in the instance group when launching nodes as Spot Instances, expressed in USD.

_igRequestedInstanceCount :: Maybe Integer

The target number of instances for the instance group.

_igRunningInstanceCount :: Maybe Integer

The number of instances currently running in this instance group.

_igInstanceGroupType :: Maybe InstanceGroupType

The type of the instance group. Valid values are MASTER, CORE or TASK.

_igInstanceType :: Maybe Text

The EC2 instance type for all instances in the instance group.

_igMarket :: Maybe MarketType

The marketplace to provision instances for this group. Valid values are ON_DEMAND or SPOT.

_igName :: Maybe Text

The name of the instance group.

_igId :: Maybe Text

The identifier of the instance group.

data InstanceGroupConfig

Configuration defining a new instance group.

Constructors

InstanceGroupConfig

Fields

_igcBidPrice :: Maybe Text

Bid price for each Amazon EC2 instance in the instance group when launching nodes as Spot Instances, expressed in USD.

_igcInstanceCount :: Integer

Target number of instances for the instance group.

_igcInstanceRole :: InstanceRoleType

The role of the instance group in the cluster.

_igcInstanceType :: Text

The Amazon EC2 instance type for all instances in the instance group.

_igcMarket :: Maybe MarketType

Market type of the Amazon EC2 instances used to create a cluster node.

_igcName :: Maybe Text

Friendly name given to the instance group.

data InstanceGroupDetail

Detailed information about an instance group.

Constructors

InstanceGroupDetail

Fields

_igdState :: InstanceGroupState

State of instance group. The following values are deprecated: STARTING, TERMINATED, and FAILED.

_igdBidPrice :: Maybe Text

Bid price for EC2 Instances when launching nodes as Spot Instances, expressed in USD.

_igdReadyDateTime :: Maybe POSIX

The date/time the instance group was available to the cluster.

_igdInstanceRole :: InstanceRoleType

Instance group role in the cluster.

_igdInstanceType :: Text

Amazon EC2 Instance type.

_igdMarket :: MarketType

Market type of the Amazon EC2 instances used to create a cluster node.

_igdLastStateChangeReason :: Maybe Text

Details regarding the state of the instance group.

_igdName :: Maybe Text

Friendly name for the instance group.

_igdInstanceRunningCount :: Integer

Actual count of running instances.

_igdCreationDateTime :: POSIX

The date/time the instance group was created.

_igdInstanceGroupId :: Maybe Text

Unique identifier for the instance group.

_igdInstanceRequestCount :: Integer

Target number of instances to run in the instance group.

_igdEndDateTime :: Maybe POSIX

The date/time the instance group was terminated.

_igdStartDateTime :: Maybe POSIX

The date/time the instance group was started.

data InstanceGroupModifyConfig

Modify an instance group size.

Constructors

InstanceGroupModifyConfig

Fields

_igmcInstanceCount :: Maybe Integer

Target size for the instance group.

_igmcEC2InstanceIdsToTerminate :: [Text]

The EC2 InstanceIds to terminate. For advanced users only. Once you terminate the instances, the instance group will not return to its original requested size.

_igmcInstanceGroupId :: Text

Unique ID of the instance group to expand or shrink.

data InstanceGroupStateChangeReason

The status change reason details for the instance group.

Constructors

InstanceGroupStateChangeReason

Fields

_igscrCode :: Maybe InstanceGroupStateChangeReasonCode

The programmable code for the state change reason.

_igscrMessage :: Maybe Text

The status change reason description.

data InstanceGroupStatus

The current status of the instance group.

Constructors

InstanceGroupStatus

Fields

_igtState :: Maybe InstanceGroupState

The current state of the instance group.

_igtStateChangeReason :: Maybe InstanceGroupStateChangeReason

The status change reason details for the instance group.

_igtTimeline :: Maybe InstanceGroupTimeline

The timeline of the instance group status over time.

data InstanceGroupTimeline

The timeline of the instance group status over time.

Constructors

InstanceGroupTimeline

Fields

_igvReadyDateTime :: Maybe POSIX

The date and time when the instance group became ready to perform tasks.

_igvCreationDateTime :: Maybe POSIX

The creation date and time of the instance group.

_igvEndDateTime :: Maybe POSIX

The date and time when the instance group terminated.

data InstanceStateChangeReason

The details of the status change reason for the instance.

Constructors

InstanceStateChangeReason

Fields

_iscrCode :: Maybe InstanceStateChangeReasonCode

The programmable code for the state change reason.

_iscrMessage :: Maybe Text

The status change reason description.

data InstanceStatus

The current status of the instance.

Constructors

InstanceStatus

Fields

_iiwState :: Maybe InstanceState

The current state of the instance.

_iiwStateChangeReason :: Maybe InstanceStateChangeReason

The details of the status change reason for the instance.

_iiwTimeline :: Maybe InstanceTimeline

The timeline of the instance status over time.

data InstanceTimeline

The timeline of the instance status over time.

Constructors

InstanceTimeline

Fields

_iihReadyDateTime :: Maybe POSIX

The date and time when the instance was ready to perform tasks.

_iihCreationDateTime :: Maybe POSIX

The creation date and time of the instance.

_iihEndDateTime :: Maybe POSIX

The date and time when the instance was terminated.

data JobFlowDetail

A description of a job flow.

Constructors

JobFlowDetail

Fields

_jfdAmiVersion :: Maybe Text

The version of the AMI used to initialize Amazon EC2 instances in the job flow. For a list of AMI versions currently supported by Amazon ElasticMapReduce, go to AMI Versions Supported in Elastic MapReduce in the Amazon Elastic MapReduce Developer's Guide.

_jfdExecutionStatusDetail :: JobFlowExecutionStatusDetail

Describes the execution status of the job flow.

_jfdJobFlowId :: Text

The job flow identifier.

_jfdSteps :: [StepDetail]

A list of steps run by the job flow.

_jfdJobFlowRole :: Maybe Text

The IAM role that was specified when the job flow was launched. The EC2 instances of the job flow assume this role.

_jfdBootstrapActions :: [BootstrapActionDetail]

A list of the bootstrap actions run by the job flow.

_jfdName :: Text

The name of the job flow.

_jfdLogUri :: Maybe Text

The location in Amazon S3 where log files for the job are stored.

_jfdInstances :: JobFlowInstancesDetail

Describes the Amazon EC2 instances of the job flow.

_jfdVisibleToAllUsers :: Maybe Bool

Specifies whether the job flow is visible to all IAM users of the AWS account associated with the job flow. If this value is set to true, all IAM users of that AWS account can view and (if they have the proper policy permissions set) manage the job flow. If it is set to false, only the IAM user that created the job flow can view and manage it. This value can be changed using the SetVisibleToAllUsers action.

_jfdSupportedProducts :: [Text]

A list of strings set by third party software when the job flow is launched. If you are not using third party software to manage the job flow this value is empty.

_jfdServiceRole :: Maybe Text

The IAM role that was specified when the job flow was launched. Amazon ElasticMapReduce will assume this role to work with AWS resources on your behalf.

data JobFlowExecutionStatusDetail

Describes the execution status of the job flow.

Constructors

JobFlowExecutionStatusDetail

Fields

_jfesdState :: JobFlowExecutionState

The state of the job flow.

_jfesdReadyDateTime :: Maybe POSIX

The date and time when the job flow was ready to start running bootstrap actions.

_jfesdLastStateChangeReason :: Maybe Text

Description of the job flow last changed state.

_jfesdCreationDateTime :: POSIX

The creation date and time of the job flow.

_jfesdEndDateTime :: Maybe POSIX

The completion date and time of the job flow.

_jfesdStartDateTime :: Maybe POSIX

The start date and time of the job flow.

data JobFlowInstancesConfig

A specification of the number and type of Amazon EC2 instances on which to run the job flow.

Constructors

JobFlowInstancesConfig

Fields

_jficEc2KeyName :: Maybe Text

The name of the Amazon EC2 key pair that can be used to ssh to the master node as the user called "hadoop.".

_jficSlaveInstanceType :: Maybe Text

The EC2 instance type of the slave nodes.

_jficInstanceCount :: Maybe Integer

The number of Amazon EC2 instances used to execute the job flow.

_jficHadoopVersion :: Maybe Text

The Hadoop version for the job flow. Valid inputs are "0.18", "0.20", or "0.20.205". If you do not set this value, the default of 0.18 is used, unless the AmiVersion parameter is set in the RunJobFlow call, in which case the default version of Hadoop for that AMI version is used.

_jficEc2SubnetId :: Maybe Text

To launch the job flow in Amazon Virtual Private Cloud (Amazon VPC), set this parameter to the identifier of the Amazon VPC subnet where you want the job flow to launch. If you do not specify this value, the job flow is launched in the normal Amazon Web Services cloud, outside of an Amazon VPC. Amazon VPC currently does not support cluster compute quadruple extra large (cc1.4xlarge) instances. Thus you cannot specify the cc1.4xlarge instance type for nodes of a job flow launched in a Amazon VPC.

_jficMasterInstanceType :: Maybe Text

The EC2 instance type of the master node.

_jficInstanceGroups :: [InstanceGroupConfig]

Configuration for the job flow's instance groups.

_jficKeepJobFlowAliveWhenNoSteps :: Maybe Bool

Specifies whether the job flow should terminate after completing all steps.

_jficTerminationProtected :: Maybe Bool

Specifies whether to lock the job flow to prevent the Amazon EC2 instances from being terminated by API call, user intervention, or in the event of a job flow error.

_jficPlacement :: Maybe PlacementType

The Availability Zone the job flow will run in.

data JobFlowInstancesDetail

Describes the Amazon EC2 instances of the job flow.

Constructors

JobFlowInstancesDetail

Fields

_jfidEc2KeyName :: Maybe Text

The name of an Amazon EC2 key pair that can be used to ssh to the master node of job flow.

_jfidSlaveInstanceType :: Text

The Amazon EC2 slave node instance type.

_jfidInstanceCount :: Integer

The number of Amazon EC2 instances in the cluster. If the value is 1, the same instance serves as both the master and slave node. If the value is greater than 1, one instance is the master node and all others are slave nodes.

_jfidNormalizedInstanceHours :: Maybe Integer

An approximation of the cost of the job flow, represented in m1.small/hours. This value is incremented once for every hour an m1.small runs. Larger instances are weighted more, so an Amazon EC2 instance that is roughly four times more expensive would result in the normalized instance hours being incremented by four. This result is only an approximation and does not reflect the actual billing rate.

_jfidHadoopVersion :: Maybe Text

The Hadoop version for the job flow.

_jfidEc2SubnetId :: Maybe Text

For job flows launched within Amazon Virtual Private Cloud, this value specifies the identifier of the subnet where the job flow was launched.

_jfidMasterInstanceType :: Text

The Amazon EC2 master node instance type.

_jfidInstanceGroups :: [InstanceGroupDetail]

Details about the job flow's instance groups.

_jfidKeepJobFlowAliveWhenNoSteps :: Maybe Bool

Specifies whether the job flow should terminate after completing all steps.

_jfidMasterInstanceId :: Maybe Text

The Amazon EC2 instance identifier of the master node.

_jfidMasterPublicDnsName :: Maybe Text

The DNS name of the master node.

_jfidTerminationProtected :: Maybe Bool

Specifies whether the Amazon EC2 instances in the cluster are protected from termination by API calls, user intervention, or in the event of a job flow error.

_jfidPlacement :: Maybe PlacementType

The Amazon EC2 Availability Zone for the job flow.

data KeyValue

A key value pair.

Constructors

KeyValue

Fields

_kvValue :: Maybe Text

The value part of the identified key.

_kvKey :: Maybe Text

The unique identifier of a key value pair.

data ScriptBootstrapActionConfig

The script run by the bootstrap action.

Constructors

ScriptBootstrapActionConfig

Fields

_sbacArgs :: [Text]

A list of command line arguments to pass to the bootstrap action script.

_sbacPath :: Text

Location of the script to run during a bootstrap action. Can be either a location in Amazon S3 or on a local file system.

data Step

The step details for the requested step identifier.

Constructors

Step

Fields

_vStatus :: Maybe StepStatus

The current execution status details of the cluster step.

_vActionOnFailure :: Maybe ActionOnFailure

This specifies what action to take when the cluster step fails. Possible values are TERMINATE_CLUSTER, CANCEL_AND_WAIT, and CONTINUE.

_vConfig :: Maybe HadoopStepConfig

The Hadoop job configuration of the cluster step.

_vName :: Maybe Text

The name of the cluster step.

_vId :: Maybe Text

The identifier of the cluster step.

data StepConfig

Specification of a job flow step.

Constructors

StepConfig

Fields

_scActionOnFailure :: Maybe ActionOnFailure

The action to take if the job flow step fails.

_scHadoopJarStep :: HadoopJarStepConfig

The JAR file used for the job flow step.

_scName :: Text

The name of the job flow step.

data StepDetail

Combines the execution state and configuration of a step.

Constructors

StepDetail

Fields

_seExecutionStatusDetail :: StepExecutionStatusDetail

The description of the step status.

_seStepConfig :: StepConfig

The step configuration.

data StepExecutionStatusDetail

The description of the step status.

Constructors

StepExecutionStatusDetail

Fields

_sesdState :: StepExecutionState

The state of the job flow step.

_sesdLastStateChangeReason :: Maybe Text

A description of the step's current state.

_sesdCreationDateTime :: POSIX

The creation date and time of the step.

_sesdEndDateTime :: Maybe POSIX

The completion date and time of the step.

_sesdStartDateTime :: Maybe POSIX

The start date and time of the step.

data StepStateChangeReason

The reason for the step execution status change.

Constructors

StepStateChangeReason

Fields

_sscrCode :: Maybe StepStateChangeReasonCode

The programmable code for the state change reason.

_sscrMessage :: Maybe Text

The descriptive message for the state change reason.

data StepStatus

The current execution status details of the cluster step.

Constructors

StepStatus

Fields

_ssState :: Maybe StepState

The execution state of the cluster step.

_ssStateChangeReason :: Maybe StepStateChangeReason

The reason for the step execution status change.

_ssTimeline :: Maybe StepTimeline

The timeline of the cluster step status over time.

data StepSummary

The summary of the cluster step.

Constructors

StepSummary

Fields

_ssyStatus :: Maybe StepStatus

The current execution status details of the cluster step.

_ssyName :: Maybe Text

The name of the cluster step.

_ssyId :: Maybe Text

The identifier of the cluster step.

data StepTimeline

The timeline of the cluster step status over time.

Constructors

StepTimeline

Fields

_suCreationDateTime :: Maybe POSIX

The date and time when the cluster step was created.

_suEndDateTime :: Maybe POSIX

The date and time when the cluster step execution completed or failed.

_suStartDateTime :: Maybe POSIX

The date and time when the cluster step execution started.

data SupportedProductConfig

The list of supported product configurations which allow user-supplied arguments. EMR accepts these arguments and forwards them to the corresponding installation script as bootstrap action arguments.

Constructors

SupportedProductConfig

Fields

_spcArgs :: [Text]

The list of user-supplied arguments.

_spcName :: Maybe Text

The name of the product configuration.

data Tag

A key/value pair containing user-defined metadata that you can associate with an Amazon EMR resource. Tags make it easier to associate clusters in various ways, such as grouping clu sters to track your Amazon EMR resource allocation costs. For more information, see Tagging Amazon EMR Resources.

Constructors

Tag

Fields

_tValue :: Maybe Text

A user-defined value, which is optional in a tag. For more information, see Tagging Amazon EMR Resources.

_tKey :: Maybe Text

A user-defined key, which is the minimum required information for a valid tag. For more information, see Tagging Amazon EMR Resources.

badBootstrapActionConfig :: Iso' BootstrapActionDetail (Maybe BootstrapActionConfig)

bootstrapActionDetail :: Iso' (Maybe BootstrapActionConfig) BootstrapActionDetail

ptAvailabilityZone :: Iso' PlacementType Text

placementType :: Iso' Text PlacementType

rVersion :: Lens' Application (Maybe Text)

rName :: Lens' Application (Maybe Text)

rArgs :: Lens' Application [Text]

rAdditionalInfo :: Lens' Application (HashMap Text Text)

bacScriptBootstrapAction :: Lens' BootstrapActionConfig ScriptBootstrapActionConfig

bacName :: Lens' BootstrapActionConfig Text

kVisibleToAllUsers :: Lens' Cluster (Maybe Bool)

kTerminationProtected :: Lens' Cluster (Maybe Bool)

kTags :: Lens' Cluster [Tag]

kStatus :: Lens' Cluster (Maybe ClusterStatus)

kServiceRole :: Lens' Cluster (Maybe Text)

kRunningAmiVersion :: Lens' Cluster (Maybe Text)

kRequestedAmiVersion :: Lens' Cluster (Maybe Text)

kName :: Lens' Cluster (Maybe Text)

kLogUri :: Lens' Cluster (Maybe Text)

kId :: Lens' Cluster (Maybe Text)

kEc2InstanceAttributes :: Lens' Cluster (Maybe Ec2InstanceAttributes)

kAutoTerminate :: Lens' Cluster (Maybe Bool)

kApplications :: Lens' Cluster [Application]

cscrMessage :: Lens' ClusterStateChangeReason (Maybe Text)

cscrCode :: Lens' ClusterStateChangeReason (Maybe ClusterStateChangeReasonCode)

csTimeline :: Lens' ClusterStatus (Maybe ClusterTimeline)

csStateChangeReason :: Lens' ClusterStatus (Maybe ClusterStateChangeReason)

csState :: Lens' ClusterStatus (Maybe ClusterState)

cwStatus :: Lens' ClusterSummary (Maybe ClusterStatus)

cwName :: Lens' ClusterSummary (Maybe Text)

cwId :: Lens' ClusterSummary (Maybe Text)

cuReadyDateTime :: Lens' ClusterTimeline (Maybe POSIX)

cuEndDateTime :: Lens' ClusterTimeline (Maybe POSIX)

cuCreationDateTime :: Lens' ClusterTimeline (Maybe POSIX)

cdScriptPath :: Lens' Command (Maybe Text)

cdName :: Lens' Command (Maybe Text)

cdArgs :: Lens' Command [Text]

eiaIamInstanceProfile :: Lens' Ec2InstanceAttributes (Maybe Text)

eiaEc2SubnetId :: Lens' Ec2InstanceAttributes (Maybe Text)

eiaEc2KeyName :: Lens' Ec2InstanceAttributes (Maybe Text)

eiaEc2AvailabilityZone :: Lens' Ec2InstanceAttributes (Maybe Text)

hjscProperties :: Lens' HadoopJarStepConfig [KeyValue]

hjscMainClass :: Lens' HadoopJarStepConfig (Maybe Text)

hjscJar :: Lens' HadoopJarStepConfig Text

hjscArgs :: Lens' HadoopJarStepConfig [Text]

hscProperties :: Lens' HadoopStepConfig (HashMap Text Text)

hscMainClass :: Lens' HadoopStepConfig (Maybe Text)

hscJar :: Lens' HadoopStepConfig (Maybe Text)

hscArgs :: Lens' HadoopStepConfig [Text]

ieStatus :: Lens' Instance (Maybe InstanceStatus)

iePublicIpAddress :: Lens' Instance (Maybe Text)

iePublicDnsName :: Lens' Instance (Maybe Text)

iePrivateIpAddress :: Lens' Instance (Maybe Text)

iePrivateDnsName :: Lens' Instance (Maybe Text)

ieId :: Lens' Instance (Maybe Text)

ieEc2InstanceId :: Lens' Instance (Maybe Text)

igStatus :: Lens' InstanceGroup (Maybe InstanceGroupStatus)

igRunningInstanceCount :: Lens' InstanceGroup (Maybe Integer)

igRequestedInstanceCount :: Lens' InstanceGroup (Maybe Integer)

igName :: Lens' InstanceGroup (Maybe Text)

igMarket :: Lens' InstanceGroup (Maybe MarketType)

igInstanceType :: Lens' InstanceGroup (Maybe Text)

igInstanceGroupType :: Lens' InstanceGroup (Maybe InstanceGroupType)

igId :: Lens' InstanceGroup (Maybe Text)

igBidPrice :: Lens' InstanceGroup (Maybe Text)

igcName :: Lens' InstanceGroupConfig (Maybe Text)

igcMarket :: Lens' InstanceGroupConfig (Maybe MarketType)

igcInstanceType :: Lens' InstanceGroupConfig Text

igcInstanceRole :: Lens' InstanceGroupConfig InstanceRoleType

igcInstanceCount :: Lens' InstanceGroupConfig Integer

igcBidPrice :: Lens' InstanceGroupConfig (Maybe Text)

igdState :: Lens' InstanceGroupDetail InstanceGroupState

igdStartDateTime :: Lens' InstanceGroupDetail (Maybe POSIX)

igdReadyDateTime :: Lens' InstanceGroupDetail (Maybe POSIX)

igdName :: Lens' InstanceGroupDetail (Maybe Text)

igdMarket :: Lens' InstanceGroupDetail MarketType

igdLastStateChangeReason :: Lens' InstanceGroupDetail (Maybe Text)

igdInstanceType :: Lens' InstanceGroupDetail Text

igdInstanceRunningCount :: Lens' InstanceGroupDetail Integer

igdInstanceRole :: Lens' InstanceGroupDetail InstanceRoleType

igdInstanceRequestCount :: Lens' InstanceGroupDetail Integer

igdInstanceGroupId :: Lens' InstanceGroupDetail (Maybe Text)

igdEndDateTime :: Lens' InstanceGroupDetail (Maybe POSIX)

igdCreationDateTime :: Lens' InstanceGroupDetail POSIX

igdBidPrice :: Lens' InstanceGroupDetail (Maybe Text)

igmcInstanceGroupId :: Lens' InstanceGroupModifyConfig Text

igmcInstanceCount :: Lens' InstanceGroupModifyConfig (Maybe Integer)

igmcEC2InstanceIdsToTerminate :: Lens' InstanceGroupModifyConfig [Text]

igscrMessage :: Lens' InstanceGroupStateChangeReason (Maybe Text)

igscrCode :: Lens' InstanceGroupStateChangeReason (Maybe InstanceGroupStateChangeReasonCode)

igtTimeline :: Lens' InstanceGroupStatus (Maybe InstanceGroupTimeline)

igtStateChangeReason :: Lens' InstanceGroupStatus (Maybe InstanceGroupStateChangeReason)

igtState :: Lens' InstanceGroupStatus (Maybe InstanceGroupState)

igvReadyDateTime :: Lens' InstanceGroupTimeline (Maybe POSIX)

igvEndDateTime :: Lens' InstanceGroupTimeline (Maybe POSIX)

igvCreationDateTime :: Lens' InstanceGroupTimeline (Maybe POSIX)

iscrMessage :: Lens' InstanceStateChangeReason (Maybe Text)

iscrCode :: Lens' InstanceStateChangeReason (Maybe InstanceStateChangeReasonCode)

iiwTimeline :: Lens' InstanceStatus (Maybe InstanceTimeline)

iiwStateChangeReason :: Lens' InstanceStatus (Maybe InstanceStateChangeReason)

iiwState :: Lens' InstanceStatus (Maybe InstanceState)

iihReadyDateTime :: Lens' InstanceTimeline (Maybe POSIX)

iihEndDateTime :: Lens' InstanceTimeline (Maybe POSIX)

iihCreationDateTime :: Lens' InstanceTimeline (Maybe POSIX)

jfdVisibleToAllUsers :: Lens' JobFlowDetail (Maybe Bool)

jfdSupportedProducts :: Lens' JobFlowDetail [Text]

jfdSteps :: Lens' JobFlowDetail [StepDetail]

jfdServiceRole :: Lens' JobFlowDetail (Maybe Text)

jfdName :: Lens' JobFlowDetail Text

jfdLogUri :: Lens' JobFlowDetail (Maybe Text)

jfdJobFlowRole :: Lens' JobFlowDetail (Maybe Text)

jfdJobFlowId :: Lens' JobFlowDetail Text

jfdInstances :: Lens' JobFlowDetail JobFlowInstancesDetail

jfdExecutionStatusDetail :: Lens' JobFlowDetail JobFlowExecutionStatusDetail

jfdBootstrapActions :: Lens' JobFlowDetail [BootstrapActionDetail]

jfdAmiVersion :: Lens' JobFlowDetail (Maybe Text)

jfesdState :: Lens' JobFlowExecutionStatusDetail JobFlowExecutionState

jfesdStartDateTime :: Lens' JobFlowExecutionStatusDetail (Maybe POSIX)

jfesdReadyDateTime :: Lens' JobFlowExecutionStatusDetail (Maybe POSIX)

jfesdLastStateChangeReason :: Lens' JobFlowExecutionStatusDetail (Maybe Text)

jfesdEndDateTime :: Lens' JobFlowExecutionStatusDetail (Maybe POSIX)

jfesdCreationDateTime :: Lens' JobFlowExecutionStatusDetail POSIX

jficTerminationProtected :: Lens' JobFlowInstancesConfig (Maybe Bool)

jficSlaveInstanceType :: Lens' JobFlowInstancesConfig (Maybe Text)

jficPlacement :: Lens' JobFlowInstancesConfig (Maybe PlacementType)

jficMasterInstanceType :: Lens' JobFlowInstancesConfig (Maybe Text)

jficKeepJobFlowAliveWhenNoSteps :: Lens' JobFlowInstancesConfig (Maybe Bool)

jficInstanceGroups :: Lens' JobFlowInstancesConfig [InstanceGroupConfig]

jficInstanceCount :: Lens' JobFlowInstancesConfig (Maybe Integer)

jficHadoopVersion :: Lens' JobFlowInstancesConfig (Maybe Text)

jficEc2SubnetId :: Lens' JobFlowInstancesConfig (Maybe Text)

jficEc2KeyName :: Lens' JobFlowInstancesConfig (Maybe Text)

jfidTerminationProtected :: Lens' JobFlowInstancesDetail (Maybe Bool)

jfidSlaveInstanceType :: Lens' JobFlowInstancesDetail Text

jfidPlacement :: Lens' JobFlowInstancesDetail (Maybe PlacementType)

jfidNormalizedInstanceHours :: Lens' JobFlowInstancesDetail (Maybe Integer)

jfidMasterPublicDnsName :: Lens' JobFlowInstancesDetail (Maybe Text)

jfidMasterInstanceType :: Lens' JobFlowInstancesDetail Text

jfidMasterInstanceId :: Lens' JobFlowInstancesDetail (Maybe Text)

jfidKeepJobFlowAliveWhenNoSteps :: Lens' JobFlowInstancesDetail (Maybe Bool)

jfidInstanceGroups :: Lens' JobFlowInstancesDetail [InstanceGroupDetail]

jfidInstanceCount :: Lens' JobFlowInstancesDetail Integer

jfidHadoopVersion :: Lens' JobFlowInstancesDetail (Maybe Text)

jfidEc2SubnetId :: Lens' JobFlowInstancesDetail (Maybe Text)

jfidEc2KeyName :: Lens' JobFlowInstancesDetail (Maybe Text)

kvValue :: Lens' KeyValue (Maybe Text)

kvKey :: Lens' KeyValue (Maybe Text)

sbacPath :: Lens' ScriptBootstrapActionConfig Text

sbacArgs :: Lens' ScriptBootstrapActionConfig [Text]

vStatus :: Lens' Step (Maybe StepStatus)

vName :: Lens' Step (Maybe Text)

vId :: Lens' Step (Maybe Text)

vConfig :: Lens' Step (Maybe HadoopStepConfig)

vActionOnFailure :: Lens' Step (Maybe ActionOnFailure)

scName :: Lens' StepConfig Text

scHadoopJarStep :: Lens' StepConfig HadoopJarStepConfig

scActionOnFailure :: Lens' StepConfig (Maybe ActionOnFailure)

seStepConfig :: Lens' StepDetail StepConfig

seExecutionStatusDetail :: Lens' StepDetail StepExecutionStatusDetail

sesdState :: Lens' StepExecutionStatusDetail StepExecutionState

sesdStartDateTime :: Lens' StepExecutionStatusDetail (Maybe POSIX)

sesdLastStateChangeReason :: Lens' StepExecutionStatusDetail (Maybe Text)

sesdEndDateTime :: Lens' StepExecutionStatusDetail (Maybe POSIX)

sesdCreationDateTime :: Lens' StepExecutionStatusDetail POSIX

sscrMessage :: Lens' StepStateChangeReason (Maybe Text)

sscrCode :: Lens' StepStateChangeReason (Maybe StepStateChangeReasonCode)

ssTimeline :: Lens' StepStatus (Maybe StepTimeline)

ssStateChangeReason :: Lens' StepStatus (Maybe StepStateChangeReason)

ssState :: Lens' StepStatus (Maybe StepState)

ssyStatus :: Lens' StepSummary (Maybe StepStatus)

ssyName :: Lens' StepSummary (Maybe Text)

ssyId :: Lens' StepSummary (Maybe Text)

suStartDateTime :: Lens' StepTimeline (Maybe POSIX)

suEndDateTime :: Lens' StepTimeline (Maybe POSIX)

suCreationDateTime :: Lens' StepTimeline (Maybe POSIX)

spcName :: Lens' SupportedProductConfig (Maybe Text)

spcArgs :: Lens' SupportedProductConfig [Text]

tValue :: Lens' Tag (Maybe Text)

tKey :: Lens' Tag (Maybe Text)