Network.AWS.EMR.V2009_03_31.ListBootstrapActions

Description

Provides information about the bootstrap actions associated with a cluster.

Documentation

listBootstrapActions

Arguments

:: Text
_lbaiClusterId

-> ListBootstrapActions

Minimum specification for a ListBootstrapActions request.

data ListBootstrapActions

Constructors

ListBootstrapActions

Fields

_lbaiClusterId :: Text

The cluster identifier for the bootstrap actions to list .

_lbaiMarker :: Maybe Text

The pagination token that indicates the next set of results to retrieve .

lbaiMarker :: Lens' ListBootstrapActions (Maybe Text)

lbaiClusterId :: Lens' ListBootstrapActions Text

data ListBootstrapActionsResponse

Constructors

ListBootstrapActionsResponse

Fields

_lbaoBootstrapActions :: [Command]

The bootstrap actions associated with the cluster .

_lbaoMarker :: Maybe Text

The pagination token that indicates the next set of results to retrieve .

lbaoMarker :: Lens' ListBootstrapActionsResponse (Maybe Text)

lbaoBootstrapActions :: Lens' ListBootstrapActionsResponse [Command]