Network.AWS.EMR.V2009_03_31.ListClusters

Description

Provides the status of all clusters visible to this AWS account. Allows you to filter the list of clusters based on certain criteria; for example, filtering by cluster creation date and time or by status. This call returns a maximum of 50 clusters per call, but returns a marker to track the paging of the cluster list across multiple ListClusters calls.

Documentation

listClusters :: ListClusters

Minimum specification for a ListClusters request.

data ListClusters

Constructors

ListClusters

Fields

_lciClusterStates :: [ClusterState]

The cluster state filters to apply when listing clusters.

_lciCreatedAfter :: Maybe POSIX

The creation date and time beginning value filter for listing clusters .

_lciCreatedBefore :: Maybe POSIX

The creation date and time end value filter for listing clusters .

_lciMarker :: Maybe Text

The pagination token that indicates the next set of results to retrieve.

lciMarker :: Lens' ListClusters (Maybe Text)

lciCreatedBefore :: Lens' ListClusters (Maybe POSIX)

lciCreatedAfter :: Lens' ListClusters (Maybe POSIX)

lciClusterStates :: Lens' ListClusters [ClusterState]

data ListClustersResponse

Constructors

ListClustersResponse

Fields

_lcoClusters :: [ClusterSummary]

The list of clusters for the account based on the given filters.

_lcoMarker :: Maybe Text

The pagination token that indicates the next set of results to retrieve.

lcoMarker :: Lens' ListClustersResponse (Maybe Text)

lcoClusters :: Lens' ListClustersResponse [ClusterSummary]