Network.AWS.EMR.V2009_03_31.ListInstanceGroups

Description

Provides all available details about the instance groups in a cluster.

Documentation

listInstanceGroups

Arguments

:: Text
_ligiClusterId

-> ListInstanceGroups

Minimum specification for a ListInstanceGroups request.

data ListInstanceGroups

Constructors

ListInstanceGroups

Fields

_ligiClusterId :: Text

The identifier of the cluster for which to list the instance groups.

_ligiMarker :: Maybe Text

The pagination token that indicates the next set of results to retrieve.

ligiMarker :: Lens' ListInstanceGroups (Maybe Text)

ligiClusterId :: Lens' ListInstanceGroups Text

data ListInstanceGroupsResponse

Constructors

ListInstanceGroupsResponse

Fields

_ligoInstanceGroups :: [InstanceGroup]

The list of instance groups for the cluster and given filters.

_ligoMarker :: Maybe Text

The pagination token that indicates the next set of results to retrieve.

ligoMarker :: Lens' ListInstanceGroupsResponse (Maybe Text)

ligoInstanceGroups :: Lens' ListInstanceGroupsResponse [InstanceGroup]