Network.AWS.EMR.V2009_03_31.ListSteps

Description

Provides a list of steps for the cluster.

Documentation

listSteps

Arguments

:: Text
_lsiClusterId

-> ListSteps

Minimum specification for a ListSteps request.

data ListSteps

Constructors

ListSteps

Fields

_lsiClusterId :: Text

The identifier of the cluster for which to list the steps.

_lsiMarker :: Maybe Text

The pagination token that indicates the next set of results to retrieve.

_lsiStepStates :: [StepState]

The filter to limit the step list based on certain states.

lsiStepStates :: Lens' ListSteps [StepState]

lsiMarker :: Lens' ListSteps (Maybe Text)

lsiClusterId :: Lens' ListSteps Text

data ListStepsResponse

Constructors

ListStepsResponse

Fields

_lsoMarker :: Maybe Text

The pagination token that indicates the next set of results to retrieve.

_lsoSteps :: [StepSummary]

The filtered list of steps for the cluster.

lsoSteps :: Lens' ListStepsResponse [StepSummary]

lsoMarker :: Lens' ListStepsResponse (Maybe Text)