Network.AWS.EMR.V2009_03_31.TerminateJobFlows

Description

TerminateJobFlows shuts a list of job flows down. When a job flow is shut down, any step not yet completed is canceled and the EC2 instances on which the job flow is running are stopped. Any log files not already saved are uploaded to Amazon S3 if a LogUri was specified when the job flow was created. The call to TerminateJobFlows is asynchronous. Depending on the configuration of the job flow, it may take up to 5-20 minutes for the job flow to completely terminate and release allocated resources, such as Amazon EC2 instances. POST HTTP1.1 Content-Type: application/x-amz-json-1.1 X-Amz-Target: ElasticMapReduce.TerminateJobFlows Content-Length: 33 User-Agent: aws-sdk-ruby1.9.2 ruby1.9.3 i386-mingw32 Host: us-east-1.elasticmapreduce.amazonaws.com X-Amz-Date: 20130716T211858Z X-Amz-Content-Sha256: ab64713f61e066e80a6083844b9249b6c6362d34a7ae7393047aa46d38b9e315 Authorization: AWS4-HMAC-SHA256 Credential=AKIAIOSFODNN7EXAMPLE20130716us-east-1elasticmapreduceaws4_request,

SignedHeaders=content-length;content-type;host;user-agent;x-amz-content-sha256;x-amz-date;x-amz-target, Signature=9791416eaf09f36aa753a324b0de27ff5cc7084b8548cc748487a2bcb3439d58 Accept: ** {JobFlowIds: ["j-3TS0OIYO4NFN"]} HTTP1.1 200 OK x-amzn-RequestId: 5551a7c9-ee5d-11e2-9542-25296c300ff0 Content-Type: application/x-amz-json-1.1 Content-Length: 0 Date: Tue, 16 Jul 2013 21:18:59 GMT.

Documentation

data TerminateJobFlows

Constructors

TerminateJobFlows

Fields

_tjfiJobFlowIds :: [Text]

A list of job flows to be shutdown.

tjfiJobFlowIds :: Iso' TerminateJobFlows [Text]

data TerminateJobFlowsResponse

Constructors

TerminateJobFlowsResponse