Network.AWS.EMRContainers.DescribeVirtualCluster

Description

Displays detailed information about a specified virtual cluster. Virtual cluster is a managed entity on Amazon EMR on EKS. You can create, describe, list and delete virtual clusters. They do not consume any additional resource in your system. A single virtual cluster maps to a single Kubernetes namespace. Given this relationship, you can model virtual clusters the same way you model Kubernetes namespaces to meet your requirements.

Creating a Request

data DescribeVirtualCluster

See: newDescribeVirtualCluster smart constructor.

Constructors

DescribeVirtualCluster'

Fields

id :: Text

The ID of the virtual cluster that will be described.

newDescribeVirtualCluster

Arguments

:: Text
$sel:id:DescribeVirtualCluster'

-> DescribeVirtualCluster

Create a value of DescribeVirtualCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:DescribeVirtualCluster', describeVirtualCluster_id - The ID of the virtual cluster that will be described.

Request Lenses

describeVirtualCluster_id :: Lens' DescribeVirtualCluster Text

The ID of the virtual cluster that will be described.

Destructuring the Response

data DescribeVirtualClusterResponse

See: newDescribeVirtualClusterResponse smart constructor.

Constructors

DescribeVirtualClusterResponse'

Fields

virtualCluster :: Maybe VirtualCluster

This output displays information about the specified virtual cluster.

httpStatus :: Int

The response's http status code.

newDescribeVirtualClusterResponse

Arguments

:: Int
$sel:httpStatus:DescribeVirtualClusterResponse'

-> DescribeVirtualClusterResponse

Create a value of DescribeVirtualClusterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:virtualCluster:DescribeVirtualClusterResponse', describeVirtualClusterResponse_virtualCluster - This output displays information about the specified virtual cluster.

$sel:httpStatus:DescribeVirtualClusterResponse', describeVirtualClusterResponse_httpStatus - The response's http status code.

Response Lenses

describeVirtualClusterResponse_virtualCluster :: Lens' DescribeVirtualClusterResponse (Maybe VirtualCluster)

This output displays information about the specified virtual cluster.

describeVirtualClusterResponse_httpStatus :: Lens' DescribeVirtualClusterResponse Int

The response's http status code.