Network.AWS.EMRContainers.Lens

Description

Operations

ListManagedEndpoints

listManagedEndpoints_states :: Lens' ListManagedEndpoints (Maybe [EndpointState])

The states of the managed endpoints.

listManagedEndpoints_createdAfter :: Lens' ListManagedEndpoints (Maybe UTCTime)

The date and time after which the endpoints are created.

listManagedEndpoints_types :: Lens' ListManagedEndpoints (Maybe [Text])

The types of the managed endpoints.

listManagedEndpoints_nextToken :: Lens' ListManagedEndpoints (Maybe Text)

The token for the next set of managed endpoints to return.

listManagedEndpoints_maxResults :: Lens' ListManagedEndpoints (Maybe Int)

The maximum number of managed endpoints that can be listed.

listManagedEndpoints_createdBefore :: Lens' ListManagedEndpoints (Maybe UTCTime)

The date and time before which the endpoints are created.

listManagedEndpoints_virtualClusterId :: Lens' ListManagedEndpoints Text

The ID of the virtual cluster.

listManagedEndpointsResponse_nextToken :: Lens' ListManagedEndpointsResponse (Maybe Text)

The token for the next set of endpoints to return.

listManagedEndpointsResponse_endpoints :: Lens' ListManagedEndpointsResponse (Maybe [Endpoint])

The managed endpoints to be listed.

listManagedEndpointsResponse_httpStatus :: Lens' ListManagedEndpointsResponse Int

The response's http status code.

CreateVirtualCluster

createVirtualCluster_tags :: Lens' CreateVirtualCluster (Maybe (HashMap Text Text))

The tags assigned to the virtual cluster.

createVirtualCluster_name :: Lens' CreateVirtualCluster Text

The specified name of the virtual cluster.

createVirtualCluster_containerProvider :: Lens' CreateVirtualCluster ContainerProvider

The container provider of the virtual cluster.

createVirtualCluster_clientToken :: Lens' CreateVirtualCluster Text

The client token of the virtual cluster.

createVirtualClusterResponse_arn :: Lens' CreateVirtualClusterResponse (Maybe Text)

This output contains the ARN of virtual cluster.

createVirtualClusterResponse_name :: Lens' CreateVirtualClusterResponse (Maybe Text)

This output contains the name of the virtual cluster.

createVirtualClusterResponse_id :: Lens' CreateVirtualClusterResponse (Maybe Text)

This output contains the virtual cluster ID.

createVirtualClusterResponse_httpStatus :: Lens' CreateVirtualClusterResponse Int

The response's http status code.

DeleteVirtualCluster

deleteVirtualCluster_id :: Lens' DeleteVirtualCluster Text

The ID of the virtual cluster that will be deleted.

deleteVirtualClusterResponse_id :: Lens' DeleteVirtualClusterResponse (Maybe Text)

This output contains the ID of the virtual cluster that will be deleted.

deleteVirtualClusterResponse_httpStatus :: Lens' DeleteVirtualClusterResponse Int

The response's http status code.

CreateManagedEndpoint

createManagedEndpoint_configurationOverrides :: Lens' CreateManagedEndpoint (Maybe ConfigurationOverrides)

The configuration settings that will be used to override existing configurations.

createManagedEndpoint_tags :: Lens' CreateManagedEndpoint (Maybe (HashMap Text Text))

The tags of the managed endpoint.

createManagedEndpoint_name :: Lens' CreateManagedEndpoint Text

The name of the managed endpoint.

createManagedEndpoint_virtualClusterId :: Lens' CreateManagedEndpoint Text

The ID of the virtual cluster for which a managed endpoint is created.

createManagedEndpoint_type :: Lens' CreateManagedEndpoint Text

The type of the managed endpoint.

createManagedEndpoint_releaseLabel :: Lens' CreateManagedEndpoint Text

The Amazon EMR release version.

createManagedEndpoint_executionRoleArn :: Lens' CreateManagedEndpoint Text

The ARN of the execution role.

createManagedEndpoint_certificateArn :: Lens' CreateManagedEndpoint Text

The certificate ARN of the managed endpoint.

createManagedEndpoint_clientToken :: Lens' CreateManagedEndpoint Text

The client idempotency token for this create call.

createManagedEndpointResponse_arn :: Lens' CreateManagedEndpointResponse (Maybe Text)

The output contains the ARN of the managed endpoint.

createManagedEndpointResponse_name :: Lens' CreateManagedEndpointResponse (Maybe Text)

The output contains the name of the managed endpoint.

createManagedEndpointResponse_id :: Lens' CreateManagedEndpointResponse (Maybe Text)

The output contains the ID of the managed endpoint.

createManagedEndpointResponse_virtualClusterId :: Lens' CreateManagedEndpointResponse (Maybe Text)

The output contains the ID of the virtual cluster.

createManagedEndpointResponse_httpStatus :: Lens' CreateManagedEndpointResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The ARN of tagged resources.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe (HashMap Text Text))

The tags assigned to resources.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

CancelJobRun

cancelJobRun_id :: Lens' CancelJobRun Text

The ID of the job run to cancel.

cancelJobRun_virtualClusterId :: Lens' CancelJobRun Text

The ID of the virtual cluster for which the job run will be canceled.

cancelJobRunResponse_id :: Lens' CancelJobRunResponse (Maybe Text)

The output contains the ID of the cancelled job run.

cancelJobRunResponse_virtualClusterId :: Lens' CancelJobRunResponse (Maybe Text)

The output contains the virtual cluster ID for which the job run is cancelled.

cancelJobRunResponse_httpStatus :: Lens' CancelJobRunResponse Int

The response's http status code.

DeleteManagedEndpoint

deleteManagedEndpoint_id :: Lens' DeleteManagedEndpoint Text

The ID of the managed endpoint.

deleteManagedEndpoint_virtualClusterId :: Lens' DeleteManagedEndpoint Text

The ID of the endpoint's virtual cluster.

deleteManagedEndpointResponse_id :: Lens' DeleteManagedEndpointResponse (Maybe Text)

The output displays the ID of the managed endpoint.

deleteManagedEndpointResponse_virtualClusterId :: Lens' DeleteManagedEndpointResponse (Maybe Text)

The output displays the ID of the endpoint's virtual cluster.

deleteManagedEndpointResponse_httpStatus :: Lens' DeleteManagedEndpointResponse Int

The response's http status code.

ListJobRuns

listJobRuns_states :: Lens' ListJobRuns (Maybe [JobRunState])

The states of the job run.

listJobRuns_createdAfter :: Lens' ListJobRuns (Maybe UTCTime)

The date and time after which the job runs were submitted.

listJobRuns_nextToken :: Lens' ListJobRuns (Maybe Text)

The token for the next set of job runs to return.

listJobRuns_name :: Lens' ListJobRuns (Maybe Text)

The name of the job run.

listJobRuns_maxResults :: Lens' ListJobRuns (Maybe Int)

The maximum number of job runs that can be listed.

listJobRuns_createdBefore :: Lens' ListJobRuns (Maybe UTCTime)

The date and time before which the job runs were submitted.

listJobRuns_virtualClusterId :: Lens' ListJobRuns Text

The ID of the virtual cluster for which to list the job run.

listJobRunsResponse_nextToken :: Lens' ListJobRunsResponse (Maybe Text)

This output displays the token for the next set of job runs.

listJobRunsResponse_jobRuns :: Lens' ListJobRunsResponse (Maybe [JobRun])

This output lists information about the specified job runs.

listJobRunsResponse_httpStatus :: Lens' ListJobRunsResponse Int

The response's http status code.

ListVirtualClusters

listVirtualClusters_states :: Lens' ListVirtualClusters (Maybe [VirtualClusterState])

The states of the requested virtual clusters.

listVirtualClusters_createdAfter :: Lens' ListVirtualClusters (Maybe UTCTime)

The date and time after which the virtual clusters are created.

listVirtualClusters_containerProviderType :: Lens' ListVirtualClusters (Maybe ContainerProviderType)

The container provider type of the virtual cluster. EKS is the only supported type as of now.

listVirtualClusters_nextToken :: Lens' ListVirtualClusters (Maybe Text)

The token for the next set of virtual clusters to return.

listVirtualClusters_containerProviderId :: Lens' ListVirtualClusters (Maybe Text)

The container provider ID of the virtual cluster.

listVirtualClusters_maxResults :: Lens' ListVirtualClusters (Maybe Int)

The maximum number of virtual clusters that can be listed.

listVirtualClusters_createdBefore :: Lens' ListVirtualClusters (Maybe UTCTime)

The date and time before which the virtual clusters are created.

listVirtualClustersResponse_nextToken :: Lens' ListVirtualClustersResponse (Maybe Text)

This output displays the token for the next set of virtual clusters.

listVirtualClustersResponse_virtualClusters :: Lens' ListVirtualClustersResponse (Maybe [VirtualCluster])

This output lists the specified virtual clusters.

listVirtualClustersResponse_httpStatus :: Lens' ListVirtualClustersResponse Int

The response's http status code.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The ARN of resources.

tagResource_tags :: Lens' TagResource (HashMap Text Text)

The tags assigned to resources.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

DescribeManagedEndpoint

describeManagedEndpoint_id :: Lens' DescribeManagedEndpoint Text

This output displays ID of the managed endpoint.

describeManagedEndpoint_virtualClusterId :: Lens' DescribeManagedEndpoint Text

The ID of the endpoint's virtual cluster.

describeManagedEndpointResponse_endpoint :: Lens' DescribeManagedEndpointResponse (Maybe Endpoint)

This output displays information about a managed endpoint.

describeManagedEndpointResponse_httpStatus :: Lens' DescribeManagedEndpointResponse Int

The response's http status code.

DescribeJobRun

describeJobRun_id :: Lens' DescribeJobRun Text

The ID of the job run request.

describeJobRun_virtualClusterId :: Lens' DescribeJobRun Text

The ID of the virtual cluster for which the job run is submitted.

describeJobRunResponse_jobRun :: Lens' DescribeJobRunResponse (Maybe JobRun)

The output displays information about a job run.

describeJobRunResponse_httpStatus :: Lens' DescribeJobRunResponse Int

The response's http status code.

UntagResource

untagResource_resourceArn :: Lens' UntagResource Text

The ARN of resources.

untagResource_tagKeys :: Lens' UntagResource [Text]

The tag keys of the resources.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

DescribeVirtualCluster

describeVirtualCluster_id :: Lens' DescribeVirtualCluster Text

The ID of the virtual cluster that will be described.

describeVirtualClusterResponse_virtualCluster :: Lens' DescribeVirtualClusterResponse (Maybe VirtualCluster)

This output displays information about the specified virtual cluster.

describeVirtualClusterResponse_httpStatus :: Lens' DescribeVirtualClusterResponse Int

The response's http status code.

StartJobRun

startJobRun_configurationOverrides :: Lens' StartJobRun (Maybe ConfigurationOverrides)

The configuration overrides for the job run.

startJobRun_name :: Lens' StartJobRun (Maybe Text)

The name of the job run.

startJobRun_tags :: Lens' StartJobRun (Maybe (HashMap Text Text))

The tags assigned to job runs.

startJobRun_virtualClusterId :: Lens' StartJobRun Text

The virtual cluster ID for which the job run request is submitted.

startJobRun_clientToken :: Lens' StartJobRun Text

The client idempotency token of the job run request.

startJobRun_executionRoleArn :: Lens' StartJobRun Text

The execution role ARN for the job run.

startJobRun_releaseLabel :: Lens' StartJobRun Text

The Amazon EMR release version to use for the job run.

startJobRun_jobDriver :: Lens' StartJobRun JobDriver

The job driver for the job run.

startJobRunResponse_arn :: Lens' StartJobRunResponse (Maybe Text)

This output lists the ARN of job run.

startJobRunResponse_name :: Lens' StartJobRunResponse (Maybe Text)

This output displays the name of the started job run.

startJobRunResponse_id :: Lens' StartJobRunResponse (Maybe Text)

This output displays the started job run ID.

startJobRunResponse_virtualClusterId :: Lens' StartJobRunResponse (Maybe Text)

This output displays the virtual cluster ID for which the job run was submitted.

startJobRunResponse_httpStatus :: Lens' StartJobRunResponse Int

The response's http status code.

Types

CloudWatchMonitoringConfiguration

cloudWatchMonitoringConfiguration_logStreamNamePrefix :: Lens' CloudWatchMonitoringConfiguration (Maybe Text)

The specified name prefix for log streams.

cloudWatchMonitoringConfiguration_logGroupName :: Lens' CloudWatchMonitoringConfiguration Text

The name of the log group for log publishing.

Configuration

configuration_configurations :: Lens' Configuration (Maybe [Configuration])

A list of additional configurations to apply within a configuration object.

configuration_properties :: Lens' Configuration (Maybe (HashMap Text Text))

A set of properties specified within a configuration classification.

configuration_classification :: Lens' Configuration Text

The classification within a configuration.

ConfigurationOverrides

configurationOverrides_monitoringConfiguration :: Lens' ConfigurationOverrides (Maybe MonitoringConfiguration)

The configurations for monitoring.

configurationOverrides_applicationConfiguration :: Lens' ConfigurationOverrides (Maybe [Configuration])

The configurations for the application running by the job run.

ContainerInfo

containerInfo_eksInfo :: Lens' ContainerInfo (Maybe EksInfo)

The information about the EKS cluster.

ContainerProvider

containerProvider_info :: Lens' ContainerProvider (Maybe ContainerInfo)

The information about the container cluster.

containerProvider_type :: Lens' ContainerProvider ContainerProviderType

The type of the container provider. EKS is the only supported type as of now.

containerProvider_id :: Lens' ContainerProvider Text

The ID of the container cluster.

EksInfo

eksInfo_namespace :: Lens' EksInfo (Maybe Text)

The namespaces of the EKS cluster.

Endpoint

endpoint_failureReason :: Lens' Endpoint (Maybe FailureReason)

The reasons why the endpoint has failed.

endpoint_state :: Lens' Endpoint (Maybe EndpointState)

The state of the endpoint.

endpoint_arn :: Lens' Endpoint (Maybe Text)

The ARN of the endpoint.

endpoint_createdAt :: Lens' Endpoint (Maybe UTCTime)

The date and time when the endpoint was created.

endpoint_subnetIds :: Lens' Endpoint (Maybe [Text])

The subnet IDs of the endpoint.

endpoint_stateDetails :: Lens' Endpoint (Maybe Text)

Additional details of the endpoint state.

endpoint_certificateArn :: Lens' Endpoint (Maybe Text)

The certificate ARN of the endpoint.

endpoint_executionRoleArn :: Lens' Endpoint (Maybe Text)

The execution role ARN of the endpoint.

endpoint_securityGroup :: Lens' Endpoint (Maybe Text)

The security group configuration of the endpoint.

endpoint_configurationOverrides :: Lens' Endpoint (Maybe ConfigurationOverrides)

The configuration settings that are used to override existing configurations for endpoints.

endpoint_name :: Lens' Endpoint (Maybe Text)

The name of the endpoint.

endpoint_releaseLabel :: Lens' Endpoint (Maybe Text)

The EMR release version to be used for the endpoint.

endpoint_id :: Lens' Endpoint (Maybe Text)

The ID of the endpoint.

endpoint_type :: Lens' Endpoint (Maybe Text)

The type of the endpoint.

endpoint_serverUrl :: Lens' Endpoint (Maybe Text)

The server URL of the endpoint.

endpoint_virtualClusterId :: Lens' Endpoint (Maybe Text)

The ID of the endpoint's virtual cluster.

endpoint_tags :: Lens' Endpoint (Maybe (HashMap Text Text))

The tags of the endpoint.

JobDriver

jobDriver_sparkSubmitJobDriver :: Lens' JobDriver (Maybe SparkSubmitJobDriver)

The job driver parameters specified for spark submit.

JobRun

jobRun_failureReason :: Lens' JobRun (Maybe FailureReason)

The reasons why the job run has failed.

jobRun_state :: Lens' JobRun (Maybe JobRunState)

The state of the job run.

jobRun_clientToken :: Lens' JobRun (Maybe Text)

The client token used to start a job run.

jobRun_arn :: Lens' JobRun (Maybe Text)

The ARN of job run.

jobRun_createdAt :: Lens' JobRun (Maybe UTCTime)

The date and time when the job run was created.

jobRun_stateDetails :: Lens' JobRun (Maybe Text)

Additional details of the job run state.

jobRun_createdBy :: Lens' JobRun (Maybe Text)

The user who created the job run.

jobRun_executionRoleArn :: Lens' JobRun (Maybe Text)

The execution role ARN of the job run.

jobRun_jobDriver :: Lens' JobRun (Maybe JobDriver)

Parameters of job driver for the job run.

jobRun_configurationOverrides :: Lens' JobRun (Maybe ConfigurationOverrides)

The configuration settings that are used to override default configuration.

jobRun_finishedAt :: Lens' JobRun (Maybe UTCTime)

The date and time when the job run has finished.

jobRun_name :: Lens' JobRun (Maybe Text)

The name of the job run.

jobRun_releaseLabel :: Lens' JobRun (Maybe Text)

The release version of Amazon EMR.

jobRun_id :: Lens' JobRun (Maybe Text)

The ID of the job run.

jobRun_virtualClusterId :: Lens' JobRun (Maybe Text)

The ID of the job run's virtual cluster.

jobRun_tags :: Lens' JobRun (Maybe (HashMap Text Text))

The assigned tags of the job run.

MonitoringConfiguration

monitoringConfiguration_persistentAppUI :: Lens' MonitoringConfiguration (Maybe PersistentAppUI)

Monitoring configurations for the persistent application UI.

monitoringConfiguration_s3MonitoringConfiguration :: Lens' MonitoringConfiguration (Maybe S3MonitoringConfiguration)

Amazon S3 configuration for monitoring log publishing.

monitoringConfiguration_cloudWatchMonitoringConfiguration :: Lens' MonitoringConfiguration (Maybe CloudWatchMonitoringConfiguration)

Monitoring configurations for CloudWatch.

S3MonitoringConfiguration

s3MonitoringConfiguration_logUri :: Lens' S3MonitoringConfiguration Text

Amazon S3 destination URI for log publishing.

SparkSubmitJobDriver

sparkSubmitJobDriver_sparkSubmitParameters :: Lens' SparkSubmitJobDriver (Maybe Text)

The Spark submit parameters that are used for job runs.

sparkSubmitJobDriver_entryPointArguments :: Lens' SparkSubmitJobDriver (Maybe [Text])

The arguments for job application.

sparkSubmitJobDriver_entryPoint :: Lens' SparkSubmitJobDriver Text

The entry point of job application.

VirtualCluster

virtualCluster_state :: Lens' VirtualCluster (Maybe VirtualClusterState)

The state of the virtual cluster.

virtualCluster_arn :: Lens' VirtualCluster (Maybe Text)

The ARN of the virtual cluster.

virtualCluster_createdAt :: Lens' VirtualCluster (Maybe UTCTime)

The date and time when the virtual cluster is created.

virtualCluster_name :: Lens' VirtualCluster (Maybe Text)

The name of the virtual cluster.

virtualCluster_id :: Lens' VirtualCluster (Maybe Text)

The ID of the virtual cluster.

virtualCluster_containerProvider :: Lens' VirtualCluster (Maybe ContainerProvider)

The container provider of the virtual cluster.

virtualCluster_tags :: Lens' VirtualCluster (Maybe (HashMap Text Text))

The assigned tags of the virtual cluster.