Network.AWS.EMRContainers.Types.Configuration

Description

Documentation

data Configuration

A configuration specification to be used when provisioning virtual clusters, which can include configurations for applications and software bundled with Amazon EMR on EKS. A configuration consists of a classification, properties, and optional nested configurations. A classification refers to an application-specific configuration file. Properties are the settings you want to change in that file.

See: newConfiguration smart constructor.

Constructors

Configuration'

Fields

configurations :: Maybe [Configuration]

A list of additional configurations to apply within a configuration object.

properties :: Maybe (Sensitive (HashMap Text Text))

A set of properties specified within a configuration classification.

classification :: Text

The classification within a configuration.

newConfiguration

Arguments

:: Text
$sel:classification:Configuration'

-> Configuration

Create a value of Configuration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configurations:Configuration', configuration_configurations - A list of additional configurations to apply within a configuration object.

$sel:properties:Configuration', configuration_properties - A set of properties specified within a configuration classification.

$sel:classification:Configuration', configuration_classification - The classification within a configuration.

configuration_configurations :: Lens' Configuration (Maybe [Configuration])

A list of additional configurations to apply within a configuration object.

configuration_properties :: Lens' Configuration (Maybe (HashMap Text Text))

A set of properties specified within a configuration classification.

configuration_classification :: Lens' Configuration Text

The classification within a configuration.