Network.AWS.EMRContainers.Types.Endpoint

Description

Documentation

data Endpoint

This entity represents the endpoint that is managed by Amazon EMR on EKS.

See: newEndpoint smart constructor.

Constructors

Endpoint'

Fields

failureReason :: Maybe FailureReason

The reasons why the endpoint has failed.

state :: Maybe EndpointState

The state of the endpoint.

arn :: Maybe Text

The ARN of the endpoint.

createdAt :: Maybe POSIX

The date and time when the endpoint was created.

subnetIds :: Maybe [Text]

The subnet IDs of the endpoint.

stateDetails :: Maybe Text

Additional details of the endpoint state.

certificateArn :: Maybe Text

The certificate ARN of the endpoint.

executionRoleArn :: Maybe Text

The execution role ARN of the endpoint.

securityGroup :: Maybe Text

The security group configuration of the endpoint.

configurationOverrides :: Maybe ConfigurationOverrides

The configuration settings that are used to override existing configurations for endpoints.

name :: Maybe Text

The name of the endpoint.

releaseLabel :: Maybe Text

The EMR release version to be used for the endpoint.

id :: Maybe Text

The ID of the endpoint.

type' :: Maybe Text

The type of the endpoint.

serverUrl :: Maybe Text

The server URL of the endpoint.

virtualClusterId :: Maybe Text

The ID of the endpoint's virtual cluster.

tags :: Maybe (HashMap Text Text)

The tags of the endpoint.

newEndpoint :: Endpoint

Create a value of Endpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:Endpoint', endpoint_failureReason - The reasons why the endpoint has failed.

$sel:state:Endpoint', endpoint_state - The state of the endpoint.

$sel:arn:Endpoint', endpoint_arn - The ARN of the endpoint.

$sel:createdAt:Endpoint', endpoint_createdAt - The date and time when the endpoint was created.

$sel:subnetIds:Endpoint', endpoint_subnetIds - The subnet IDs of the endpoint.

$sel:stateDetails:Endpoint', endpoint_stateDetails - Additional details of the endpoint state.

$sel:certificateArn:Endpoint', endpoint_certificateArn - The certificate ARN of the endpoint.

$sel:executionRoleArn:Endpoint', endpoint_executionRoleArn - The execution role ARN of the endpoint.

$sel:securityGroup:Endpoint', endpoint_securityGroup - The security group configuration of the endpoint.

$sel:configurationOverrides:Endpoint', endpoint_configurationOverrides - The configuration settings that are used to override existing configurations for endpoints.

$sel:name:Endpoint', endpoint_name - The name of the endpoint.

$sel:releaseLabel:Endpoint', endpoint_releaseLabel - The EMR release version to be used for the endpoint.

$sel:id:Endpoint', endpoint_id - The ID of the endpoint.

$sel:type':Endpoint', endpoint_type - The type of the endpoint.

$sel:serverUrl:Endpoint', endpoint_serverUrl - The server URL of the endpoint.

$sel:virtualClusterId:Endpoint', endpoint_virtualClusterId - The ID of the endpoint's virtual cluster.

$sel:tags:Endpoint', endpoint_tags - The tags of the endpoint.

endpoint_failureReason :: Lens' Endpoint (Maybe FailureReason)

The reasons why the endpoint has failed.

endpoint_state :: Lens' Endpoint (Maybe EndpointState)

The state of the endpoint.

endpoint_arn :: Lens' Endpoint (Maybe Text)

The ARN of the endpoint.

endpoint_createdAt :: Lens' Endpoint (Maybe UTCTime)

The date and time when the endpoint was created.

endpoint_subnetIds :: Lens' Endpoint (Maybe [Text])

The subnet IDs of the endpoint.

endpoint_stateDetails :: Lens' Endpoint (Maybe Text)

Additional details of the endpoint state.

endpoint_certificateArn :: Lens' Endpoint (Maybe Text)

The certificate ARN of the endpoint.

endpoint_executionRoleArn :: Lens' Endpoint (Maybe Text)

The execution role ARN of the endpoint.

endpoint_securityGroup :: Lens' Endpoint (Maybe Text)

The security group configuration of the endpoint.

endpoint_configurationOverrides :: Lens' Endpoint (Maybe ConfigurationOverrides)

The configuration settings that are used to override existing configurations for endpoints.

endpoint_name :: Lens' Endpoint (Maybe Text)

The name of the endpoint.

endpoint_releaseLabel :: Lens' Endpoint (Maybe Text)

The EMR release version to be used for the endpoint.

endpoint_id :: Lens' Endpoint (Maybe Text)

The ID of the endpoint.

endpoint_type :: Lens' Endpoint (Maybe Text)

The type of the endpoint.

endpoint_serverUrl :: Lens' Endpoint (Maybe Text)

The server URL of the endpoint.

endpoint_virtualClusterId :: Lens' Endpoint (Maybe Text)

The ID of the endpoint's virtual cluster.

endpoint_tags :: Lens' Endpoint (Maybe (HashMap Text Text))

The tags of the endpoint.