Network.AWS.ElastiCache.AddTagsToResource

Description

A tag is a key-value pair where the key and value are case-sensitive. You can use tags to categorize and track all your ElastiCache resources, with the exception of global replication group. When you add or remove tags on replication groups, those actions will be replicated to all nodes in the replication group. For more information, see Resource-level permissions.

For example, you can use cost-allocation tags to your ElastiCache resources, Amazon generates a cost allocation report as a comma-separated value (CSV) file with your usage and costs aggregated by your tags. You can apply tags that represent business categories (such as cost centers, application names, or owners) to organize your costs across multiple services.

For more information, see Using Cost Allocation Tags in Amazon ElastiCache in the ElastiCache User Guide.

Creating a Request

data AddTagsToResource

Represents the input of an AddTagsToResource operation.

See: newAddTagsToResource smart constructor.

Constructors

AddTagsToResource'

Fields

resourceName :: Text

The Amazon Resource Name (ARN) of the resource to which the tags are to be added, for example arn:aws:elasticache:us-west-2:0123456789:cluster:myCluster or arn:aws:elasticache:us-west-2:0123456789:snapshot:mySnapshot. ElastiCache resources are cluster and snapshot.

For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Service Namespaces.

tags :: [Tag]

A list of tags to be added to this resource. A tag is a key-value pair. A tag key must be accompanied by a tag value, although null is accepted.

newAddTagsToResource

Arguments

:: Text
$sel:resourceName:AddTagsToResource'

-> AddTagsToResource

Create a value of AddTagsToResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceName:AddTagsToResource', addTagsToResource_resourceName - The Amazon Resource Name (ARN) of the resource to which the tags are to be added, for example arn:aws:elasticache:us-west-2:0123456789:cluster:myCluster or arn:aws:elasticache:us-west-2:0123456789:snapshot:mySnapshot. ElastiCache resources are cluster and snapshot.

For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Service Namespaces.

$sel:tags:AddTagsToResource', addTagsToResource_tags - A list of tags to be added to this resource. A tag is a key-value pair. A tag key must be accompanied by a tag value, although null is accepted.

Request Lenses

addTagsToResource_resourceName :: Lens' AddTagsToResource Text

The Amazon Resource Name (ARN) of the resource to which the tags are to be added, for example arn:aws:elasticache:us-west-2:0123456789:cluster:myCluster or arn:aws:elasticache:us-west-2:0123456789:snapshot:mySnapshot. ElastiCache resources are cluster and snapshot.

For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Service Namespaces.

addTagsToResource_tags :: Lens' AddTagsToResource [Tag]

A list of tags to be added to this resource. A tag is a key-value pair. A tag key must be accompanied by a tag value, although null is accepted.

Destructuring the Response

data TagListMessage

Represents the output from the AddTagsToResource, ListTagsForResource, and RemoveTagsFromResource operations.

See: newTagListMessage smart constructor.

Constructors

TagListMessage'

Fields

tagList :: Maybe [Tag]

A list of tags as key-value pairs.

newTagListMessage :: TagListMessage

Create a value of TagListMessage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tagList:TagListMessage', tagListMessage_tagList - A list of tags as key-value pairs.

Response Lenses

tagListMessage_tagList :: Lens' TagListMessage (Maybe [Tag])

A list of tags as key-value pairs.