Network.AWS.ElastiCache.BatchStopUpdateAction

Description

Stop the service update. For more information on service updates and stopping them, see Stopping Service Updates.

Creating a Request

data BatchStopUpdateAction

See: newBatchStopUpdateAction smart constructor.

Constructors

BatchStopUpdateAction'

Fields

cacheClusterIds :: Maybe [Text]

The cache cluster IDs

replicationGroupIds :: Maybe [Text]

The replication group IDs

serviceUpdateName :: Text

The unique ID of the service update

newBatchStopUpdateAction

Arguments

:: Text
$sel:serviceUpdateName:BatchStopUpdateAction'

-> BatchStopUpdateAction

Create a value of BatchStopUpdateAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cacheClusterIds:BatchStopUpdateAction', batchStopUpdateAction_cacheClusterIds - The cache cluster IDs

$sel:replicationGroupIds:BatchStopUpdateAction', batchStopUpdateAction_replicationGroupIds - The replication group IDs

$sel:serviceUpdateName:BatchStopUpdateAction', batchStopUpdateAction_serviceUpdateName - The unique ID of the service update

Request Lenses

batchStopUpdateAction_cacheClusterIds :: Lens' BatchStopUpdateAction (Maybe [Text])

The cache cluster IDs

batchStopUpdateAction_replicationGroupIds :: Lens' BatchStopUpdateAction (Maybe [Text])

The replication group IDs

batchStopUpdateAction_serviceUpdateName :: Lens' BatchStopUpdateAction Text

The unique ID of the service update

Destructuring the Response

data UpdateActionResultsMessage

See: newUpdateActionResultsMessage smart constructor.

Constructors

UpdateActionResultsMessage'

Fields

unprocessedUpdateActions :: Maybe [UnprocessedUpdateAction]

Update actions that haven't been processed successfully

processedUpdateActions :: Maybe [ProcessedUpdateAction]

Update actions that have been processed successfully

newUpdateActionResultsMessage :: UpdateActionResultsMessage

Create a value of UpdateActionResultsMessage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:unprocessedUpdateActions:UpdateActionResultsMessage', updateActionResultsMessage_unprocessedUpdateActions - Update actions that haven't been processed successfully

$sel:processedUpdateActions:UpdateActionResultsMessage', updateActionResultsMessage_processedUpdateActions - Update actions that have been processed successfully

Response Lenses

updateActionResultsMessage_unprocessedUpdateActions :: Lens' UpdateActionResultsMessage (Maybe [UnprocessedUpdateAction])

Update actions that haven't been processed successfully

updateActionResultsMessage_processedUpdateActions :: Lens' UpdateActionResultsMessage (Maybe [ProcessedUpdateAction])

Update actions that have been processed successfully