Network.AWS.ElastiCache.DecreaseNodeGroupsInGlobalReplicationGroup

Description

Decreases the number of node groups in a Global datastore

Creating a Request

data DecreaseNodeGroupsInGlobalReplicationGroup

See: newDecreaseNodeGroupsInGlobalReplicationGroup smart constructor.

Constructors

DecreaseNodeGroupsInGlobalReplicationGroup'

Fields

globalNodeGroupsToRemove :: Maybe [Text]

If the value of NodeGroupCount is less than the current number of node groups (shards), then either NodeGroupsToRemove or NodeGroupsToRetain is required. GlobalNodeGroupsToRemove is a list of NodeGroupIds to remove from the cluster. ElastiCache for Redis will attempt to remove all node groups listed by GlobalNodeGroupsToRemove from the cluster.

globalNodeGroupsToRetain :: Maybe [Text]

If the value of NodeGroupCount is less than the current number of node groups (shards), then either NodeGroupsToRemove or NodeGroupsToRetain is required. GlobalNodeGroupsToRetain is a list of NodeGroupIds to retain from the cluster. ElastiCache for Redis will attempt to retain all node groups listed by GlobalNodeGroupsToRetain from the cluster.

globalReplicationGroupId :: Text

The name of the Global datastore

nodeGroupCount :: Int

The number of node groups (shards) that results from the modification of the shard configuration

applyImmediately :: Bool

Indicates that the shard reconfiguration process begins immediately. At present, the only permitted value for this parameter is true.

newDecreaseNodeGroupsInGlobalReplicationGroup

Arguments

:: Text
$sel:globalReplicationGroupId:DecreaseNodeGroupsInGlobalReplicationGroup'

-> Int
$sel:nodeGroupCount:DecreaseNodeGroupsInGlobalReplicationGroup'

-> Bool
$sel:applyImmediately:DecreaseNodeGroupsInGlobalReplicationGroup'

-> DecreaseNodeGroupsInGlobalReplicationGroup

Create a value of DecreaseNodeGroupsInGlobalReplicationGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:globalNodeGroupsToRemove:DecreaseNodeGroupsInGlobalReplicationGroup', decreaseNodeGroupsInGlobalReplicationGroup_globalNodeGroupsToRemove - If the value of NodeGroupCount is less than the current number of node groups (shards), then either NodeGroupsToRemove or NodeGroupsToRetain is required. GlobalNodeGroupsToRemove is a list of NodeGroupIds to remove from the cluster. ElastiCache for Redis will attempt to remove all node groups listed by GlobalNodeGroupsToRemove from the cluster.

$sel:globalNodeGroupsToRetain:DecreaseNodeGroupsInGlobalReplicationGroup', decreaseNodeGroupsInGlobalReplicationGroup_globalNodeGroupsToRetain - If the value of NodeGroupCount is less than the current number of node groups (shards), then either NodeGroupsToRemove or NodeGroupsToRetain is required. GlobalNodeGroupsToRetain is a list of NodeGroupIds to retain from the cluster. ElastiCache for Redis will attempt to retain all node groups listed by GlobalNodeGroupsToRetain from the cluster.

$sel:globalReplicationGroupId:DecreaseNodeGroupsInGlobalReplicationGroup', decreaseNodeGroupsInGlobalReplicationGroup_globalReplicationGroupId - The name of the Global datastore

$sel:nodeGroupCount:DecreaseNodeGroupsInGlobalReplicationGroup', decreaseNodeGroupsInGlobalReplicationGroup_nodeGroupCount - The number of node groups (shards) that results from the modification of the shard configuration

$sel:applyImmediately:DecreaseNodeGroupsInGlobalReplicationGroup', decreaseNodeGroupsInGlobalReplicationGroup_applyImmediately - Indicates that the shard reconfiguration process begins immediately. At present, the only permitted value for this parameter is true.

Request Lenses

decreaseNodeGroupsInGlobalReplicationGroup_globalNodeGroupsToRemove :: Lens' DecreaseNodeGroupsInGlobalReplicationGroup (Maybe [Text])

If the value of NodeGroupCount is less than the current number of node groups (shards), then either NodeGroupsToRemove or NodeGroupsToRetain is required. GlobalNodeGroupsToRemove is a list of NodeGroupIds to remove from the cluster. ElastiCache for Redis will attempt to remove all node groups listed by GlobalNodeGroupsToRemove from the cluster.

decreaseNodeGroupsInGlobalReplicationGroup_globalNodeGroupsToRetain :: Lens' DecreaseNodeGroupsInGlobalReplicationGroup (Maybe [Text])

If the value of NodeGroupCount is less than the current number of node groups (shards), then either NodeGroupsToRemove or NodeGroupsToRetain is required. GlobalNodeGroupsToRetain is a list of NodeGroupIds to retain from the cluster. ElastiCache for Redis will attempt to retain all node groups listed by GlobalNodeGroupsToRetain from the cluster.

decreaseNodeGroupsInGlobalReplicationGroup_globalReplicationGroupId :: Lens' DecreaseNodeGroupsInGlobalReplicationGroup Text

The name of the Global datastore

decreaseNodeGroupsInGlobalReplicationGroup_nodeGroupCount :: Lens' DecreaseNodeGroupsInGlobalReplicationGroup Int

The number of node groups (shards) that results from the modification of the shard configuration

decreaseNodeGroupsInGlobalReplicationGroup_applyImmediately :: Lens' DecreaseNodeGroupsInGlobalReplicationGroup Bool

Indicates that the shard reconfiguration process begins immediately. At present, the only permitted value for this parameter is true.

Destructuring the Response

data DecreaseNodeGroupsInGlobalReplicationGroupResponse

See: newDecreaseNodeGroupsInGlobalReplicationGroupResponse smart constructor.

Constructors

DecreaseNodeGroupsInGlobalReplicationGroupResponse'

Fields

globalReplicationGroup :: Maybe GlobalReplicationGroup

httpStatus :: Int

The response's http status code.

newDecreaseNodeGroupsInGlobalReplicationGroupResponse

Arguments

:: Int
$sel:httpStatus:DecreaseNodeGroupsInGlobalReplicationGroupResponse'

-> DecreaseNodeGroupsInGlobalReplicationGroupResponse

Create a value of DecreaseNodeGroupsInGlobalReplicationGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:globalReplicationGroup:DecreaseNodeGroupsInGlobalReplicationGroupResponse', decreaseNodeGroupsInGlobalReplicationGroupResponse_globalReplicationGroup - Undocumented member.

$sel:httpStatus:DecreaseNodeGroupsInGlobalReplicationGroupResponse', decreaseNodeGroupsInGlobalReplicationGroupResponse_httpStatus - The response's http status code.

Response Lenses

decreaseNodeGroupsInGlobalReplicationGroupResponse_globalReplicationGroup :: Lens' DecreaseNodeGroupsInGlobalReplicationGroupResponse (Maybe GlobalReplicationGroup)

Undocumented member.

decreaseNodeGroupsInGlobalReplicationGroupResponse_httpStatus :: Lens' DecreaseNodeGroupsInGlobalReplicationGroupResponse Int

The response's http status code.