Network.AWS.ElastiCache.DeleteCacheCluster

Description

Deletes a previously provisioned cluster. DeleteCacheCluster deletes all associated cache nodes, node endpoints and the cluster itself. When you receive a successful response from this operation, Amazon ElastiCache immediately begins deleting the cluster; you cannot cancel or revert this operation.

This operation is not valid for:

Redis (cluster mode enabled) clusters

Redis (cluster mode disabled) clusters

A cluster that is the last read replica of a replication group

A cluster that is the primary node of a replication group

A node group (shard) that has Multi-AZ mode enabled

A cluster from a Redis (cluster mode enabled) replication group

A cluster that is not in the available state

Creating a Request

data DeleteCacheCluster

Represents the input of a DeleteCacheCluster operation.

See: newDeleteCacheCluster smart constructor.

Constructors

DeleteCacheCluster'

Fields

finalSnapshotIdentifier :: Maybe Text

The user-supplied name of a final cluster snapshot. This is the unique name that identifies the snapshot. ElastiCache creates the snapshot, and then deletes the cluster immediately afterward.

cacheClusterId :: Text

The cluster identifier for the cluster to be deleted. This parameter is not case sensitive.

newDeleteCacheCluster

Arguments

:: Text
$sel:cacheClusterId:DeleteCacheCluster'

-> DeleteCacheCluster

Create a value of DeleteCacheCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:finalSnapshotIdentifier:DeleteCacheCluster', deleteCacheCluster_finalSnapshotIdentifier - The user-supplied name of a final cluster snapshot. This is the unique name that identifies the snapshot. ElastiCache creates the snapshot, and then deletes the cluster immediately afterward.

$sel:cacheClusterId:DeleteCacheCluster', deleteCacheCluster_cacheClusterId - The cluster identifier for the cluster to be deleted. This parameter is not case sensitive.

Request Lenses

deleteCacheCluster_finalSnapshotIdentifier :: Lens' DeleteCacheCluster (Maybe Text)

The user-supplied name of a final cluster snapshot. This is the unique name that identifies the snapshot. ElastiCache creates the snapshot, and then deletes the cluster immediately afterward.

deleteCacheCluster_cacheClusterId :: Lens' DeleteCacheCluster Text

The cluster identifier for the cluster to be deleted. This parameter is not case sensitive.

Destructuring the Response

data DeleteCacheClusterResponse

See: newDeleteCacheClusterResponse smart constructor.

Constructors

DeleteCacheClusterResponse'

Fields

cacheCluster :: Maybe CacheCluster

httpStatus :: Int

The response's http status code.

newDeleteCacheClusterResponse

Arguments

:: Int
$sel:httpStatus:DeleteCacheClusterResponse'

-> DeleteCacheClusterResponse

Create a value of DeleteCacheClusterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cacheCluster:DeleteCacheClusterResponse', deleteCacheClusterResponse_cacheCluster - Undocumented member.

$sel:httpStatus:DeleteCacheClusterResponse', deleteCacheClusterResponse_httpStatus - The response's http status code.

Response Lenses

deleteCacheClusterResponse_cacheCluster :: Lens' DeleteCacheClusterResponse (Maybe CacheCluster)

Undocumented member.

deleteCacheClusterResponse_httpStatus :: Lens' DeleteCacheClusterResponse Int

The response's http status code.