Network.AWS.ElastiCache.DeleteUser

Description

For Redis engine version 6.x onwards: Deletes a user. The user will be removed from all user groups and in turn removed from all replication groups. For more information, see Using Role Based Access Control (RBAC).

Creating a Request

data DeleteUser

See: newDeleteUser smart constructor.

Constructors

DeleteUser'

Fields

userId :: Text

The ID of the user.

newDeleteUser

Arguments

:: Text
$sel:userId:DeleteUser'

-> DeleteUser

Create a value of DeleteUser with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userId:DeleteUser', deleteUser_userId - The ID of the user.

Request Lenses

deleteUser_userId :: Lens' DeleteUser Text

The ID of the user.

Destructuring the Response

data User

See: newUser smart constructor.

Constructors

User'

Fields

status :: Maybe Text

Indicates the user status. Can be "active", "modifying" or "deleting".

arn :: Maybe Text

The Amazon Resource Name (ARN) of the user.

userGroupIds :: Maybe [Text]

Returns a list of the user group IDs the user belongs to.

authentication :: Maybe Authentication

Denotes whether the user requires a password to authenticate.

engine :: Maybe Text

The current supported value is Redis.

userName :: Maybe Text

The username of the user.

accessString :: Maybe Text

Access permissions string used for this user.

userId :: Maybe Text

The ID of the user.

newUser :: User

Create a value of User with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:User', user_status - Indicates the user status. Can be "active", "modifying" or "deleting".

$sel:arn:User', user_arn - The Amazon Resource Name (ARN) of the user.

$sel:userGroupIds:User', user_userGroupIds - Returns a list of the user group IDs the user belongs to.

$sel:authentication:User', user_authentication - Denotes whether the user requires a password to authenticate.

$sel:engine:User', user_engine - The current supported value is Redis.

$sel:userName:User', user_userName - The username of the user.

$sel:accessString:User', user_accessString - Access permissions string used for this user.

$sel:userId:User', user_userId - The ID of the user.

Response Lenses

user_status :: Lens' User (Maybe Text)

Indicates the user status. Can be "active", "modifying" or "deleting".

user_arn :: Lens' User (Maybe Text)

The Amazon Resource Name (ARN) of the user.

user_userGroupIds :: Lens' User (Maybe [Text])

Returns a list of the user group IDs the user belongs to.

user_authentication :: Lens' User (Maybe Authentication)

Denotes whether the user requires a password to authenticate.

user_engine :: Lens' User (Maybe Text)

The current supported value is Redis.

user_userName :: Lens' User (Maybe Text)

The username of the user.

user_accessString :: Lens' User (Maybe Text)

Access permissions string used for this user.

user_userId :: Lens' User (Maybe Text)

The ID of the user.