Network.AWS.ElastiCache.DeleteUserGroup

Description

For Redis engine version 6.x onwards: Deletes a user group. The user group must first be disassociated from the replication group before it can be deleted. For more information, see Using Role Based Access Control (RBAC).

Creating a Request

data DeleteUserGroup

See: newDeleteUserGroup smart constructor.

Constructors

DeleteUserGroup'

Fields

userGroupId :: Text

The ID of the user group.

newDeleteUserGroup

Arguments

:: Text
$sel:userGroupId:DeleteUserGroup'

-> DeleteUserGroup

Create a value of DeleteUserGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userGroupId:DeleteUserGroup', deleteUserGroup_userGroupId - The ID of the user group.

Request Lenses

deleteUserGroup_userGroupId :: Lens' DeleteUserGroup Text

The ID of the user group.

Destructuring the Response

data UserGroup

See: newUserGroup smart constructor.

Constructors

UserGroup'

Fields

status :: Maybe Text

Indicates user group status. Can be "creating", "active", "modifying", "deleting".

userIds :: Maybe [Text]

The list of user IDs that belong to the user group.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the user group.

userGroupId :: Maybe Text

The ID of the user group.

engine :: Maybe Text

The current supported value is Redis.

pendingChanges :: Maybe UserGroupPendingChanges

A list of updates being applied to the user group.

replicationGroups :: Maybe [Text]

A list of replication groups that the user group can access.

newUserGroup :: UserGroup

Create a value of UserGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:UserGroup', userGroup_status - Indicates user group status. Can be "creating", "active", "modifying", "deleting".

$sel:userIds:UserGroup', userGroup_userIds - The list of user IDs that belong to the user group.

$sel:arn:UserGroup', userGroup_arn - The Amazon Resource Name (ARN) of the user group.

$sel:userGroupId:UserGroup', userGroup_userGroupId - The ID of the user group.

$sel:engine:UserGroup', userGroup_engine - The current supported value is Redis.

$sel:pendingChanges:UserGroup', userGroup_pendingChanges - A list of updates being applied to the user group.

$sel:replicationGroups:UserGroup', userGroup_replicationGroups - A list of replication groups that the user group can access.

Response Lenses

userGroup_status :: Lens' UserGroup (Maybe Text)

Indicates user group status. Can be "creating", "active", "modifying", "deleting".

userGroup_userIds :: Lens' UserGroup (Maybe [Text])

The list of user IDs that belong to the user group.

userGroup_arn :: Lens' UserGroup (Maybe Text)

The Amazon Resource Name (ARN) of the user group.

userGroup_userGroupId :: Lens' UserGroup (Maybe Text)

The ID of the user group.

userGroup_engine :: Lens' UserGroup (Maybe Text)

The current supported value is Redis.

userGroup_pendingChanges :: Lens' UserGroup (Maybe UserGroupPendingChanges)

A list of updates being applied to the user group.

userGroup_replicationGroups :: Lens' UserGroup (Maybe [Text])

A list of replication groups that the user group can access.