Network.AWS.ElastiCache.DescribeCacheParameters

Description

Returns the detailed parameter list for a particular cache parameter group.

This operation returns paginated results.

Creating a Request

data DescribeCacheParameters

Represents the input of a DescribeCacheParameters operation.

See: newDescribeCacheParameters smart constructor.

Constructors

DescribeCacheParameters'

Fields

source :: Maybe Text

The parameter types to return.

Valid values: user | system | engine-default

maxRecords :: Maybe Int

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: minimum 20; maximum 100.

marker :: Maybe Text

An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

cacheParameterGroupName :: Text

The name of a specific cache parameter group to return details for.

newDescribeCacheParameters

Arguments

:: Text
$sel:cacheParameterGroupName:DescribeCacheParameters'

-> DescribeCacheParameters

Create a value of DescribeCacheParameters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:source:DescribeCacheParameters', describeCacheParameters_source - The parameter types to return.

Valid values: user | system | engine-default

$sel:maxRecords:DescribeCacheParameters', describeCacheParameters_maxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: minimum 20; maximum 100.

$sel:marker:DescribeCacheParameters', describeCacheParameters_marker - An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

$sel:cacheParameterGroupName:DescribeCacheParameters', describeCacheParameters_cacheParameterGroupName - The name of a specific cache parameter group to return details for.

Request Lenses

describeCacheParameters_source :: Lens' DescribeCacheParameters (Maybe Text)

The parameter types to return.

Valid values: user | system | engine-default

describeCacheParameters_maxRecords :: Lens' DescribeCacheParameters (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: minimum 20; maximum 100.

describeCacheParameters_marker :: Lens' DescribeCacheParameters (Maybe Text)

An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

describeCacheParameters_cacheParameterGroupName :: Lens' DescribeCacheParameters Text

The name of a specific cache parameter group to return details for.

Destructuring the Response

data DescribeCacheParametersResponse

Represents the output of a DescribeCacheParameters operation.

See: newDescribeCacheParametersResponse smart constructor.

Constructors

DescribeCacheParametersResponse'

Fields

cacheNodeTypeSpecificParameters :: Maybe [CacheNodeTypeSpecificParameter]

A list of parameters specific to a particular cache node type. Each element in the list contains detailed information about one parameter.

parameters :: Maybe [Parameter]

A list of Parameter instances.

marker :: Maybe Text

Provides an identifier to allow retrieval of paginated results.

httpStatus :: Int

The response's http status code.

newDescribeCacheParametersResponse

Arguments

:: Int
$sel:httpStatus:DescribeCacheParametersResponse'

-> DescribeCacheParametersResponse

Create a value of DescribeCacheParametersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cacheNodeTypeSpecificParameters:DescribeCacheParametersResponse', describeCacheParametersResponse_cacheNodeTypeSpecificParameters - A list of parameters specific to a particular cache node type. Each element in the list contains detailed information about one parameter.

$sel:parameters:DescribeCacheParametersResponse', describeCacheParametersResponse_parameters - A list of Parameter instances.

$sel:marker:DescribeCacheParameters', describeCacheParametersResponse_marker - Provides an identifier to allow retrieval of paginated results.

$sel:httpStatus:DescribeCacheParametersResponse', describeCacheParametersResponse_httpStatus - The response's http status code.

Response Lenses

describeCacheParametersResponse_cacheNodeTypeSpecificParameters :: Lens' DescribeCacheParametersResponse (Maybe [CacheNodeTypeSpecificParameter])

A list of parameters specific to a particular cache node type. Each element in the list contains detailed information about one parameter.

describeCacheParametersResponse_parameters :: Lens' DescribeCacheParametersResponse (Maybe [Parameter])

A list of Parameter instances.

describeCacheParametersResponse_marker :: Lens' DescribeCacheParametersResponse (Maybe Text)

Provides an identifier to allow retrieval of paginated results.

describeCacheParametersResponse_httpStatus :: Lens' DescribeCacheParametersResponse Int

The response's http status code.