Network.AWS.ElastiCache.DescribeCacheSecurityGroups

Description

Returns a list of cache security group descriptions. If a cache security group name is specified, the list contains only the description of that group. This applicable only when you have ElastiCache in Classic setup

This operation returns paginated results.

Creating a Request

data DescribeCacheSecurityGroups

Represents the input of a DescribeCacheSecurityGroups operation.

See: newDescribeCacheSecurityGroups smart constructor.

Constructors

DescribeCacheSecurityGroups'

Fields

cacheSecurityGroupName :: Maybe Text

The name of the cache security group to return details for.

maxRecords :: Maybe Int

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: minimum 20; maximum 100.

marker :: Maybe Text

An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

newDescribeCacheSecurityGroups :: DescribeCacheSecurityGroups

Create a value of DescribeCacheSecurityGroups with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cacheSecurityGroupName:DescribeCacheSecurityGroups', describeCacheSecurityGroups_cacheSecurityGroupName - The name of the cache security group to return details for.

$sel:maxRecords:DescribeCacheSecurityGroups', describeCacheSecurityGroups_maxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: minimum 20; maximum 100.

$sel:marker:DescribeCacheSecurityGroups', describeCacheSecurityGroups_marker - An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Request Lenses

describeCacheSecurityGroups_cacheSecurityGroupName :: Lens' DescribeCacheSecurityGroups (Maybe Text)

The name of the cache security group to return details for.

describeCacheSecurityGroups_maxRecords :: Lens' DescribeCacheSecurityGroups (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: minimum 20; maximum 100.

describeCacheSecurityGroups_marker :: Lens' DescribeCacheSecurityGroups (Maybe Text)

An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Destructuring the Response

data DescribeCacheSecurityGroupsResponse

Represents the output of a DescribeCacheSecurityGroups operation.

See: newDescribeCacheSecurityGroupsResponse smart constructor.

Constructors

DescribeCacheSecurityGroupsResponse'

Fields

cacheSecurityGroups :: Maybe [CacheSecurityGroup]

A list of cache security groups. Each element in the list contains detailed information about one group.

marker :: Maybe Text

Provides an identifier to allow retrieval of paginated results.

httpStatus :: Int

The response's http status code.

newDescribeCacheSecurityGroupsResponse

Arguments

:: Int
$sel:httpStatus:DescribeCacheSecurityGroupsResponse'

-> DescribeCacheSecurityGroupsResponse

Create a value of DescribeCacheSecurityGroupsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cacheSecurityGroups:DescribeCacheSecurityGroupsResponse', describeCacheSecurityGroupsResponse_cacheSecurityGroups - A list of cache security groups. Each element in the list contains detailed information about one group.

$sel:marker:DescribeCacheSecurityGroups', describeCacheSecurityGroupsResponse_marker - Provides an identifier to allow retrieval of paginated results.

$sel:httpStatus:DescribeCacheSecurityGroupsResponse', describeCacheSecurityGroupsResponse_httpStatus - The response's http status code.

Response Lenses

describeCacheSecurityGroupsResponse_cacheSecurityGroups :: Lens' DescribeCacheSecurityGroupsResponse (Maybe [CacheSecurityGroup])

A list of cache security groups. Each element in the list contains detailed information about one group.

describeCacheSecurityGroupsResponse_marker :: Lens' DescribeCacheSecurityGroupsResponse (Maybe Text)

Provides an identifier to allow retrieval of paginated results.

describeCacheSecurityGroupsResponse_httpStatus :: Lens' DescribeCacheSecurityGroupsResponse Int

The response's http status code.