Network.AWS.ElastiCache.DescribeCacheSubnetGroups

Description

Returns a list of cache subnet group descriptions. If a subnet group name is specified, the list contains only the description of that group. This is applicable only when you have ElastiCache in VPC setup. All ElastiCache clusters now launch in VPC by default.

This operation returns paginated results.

Creating a Request

data DescribeCacheSubnetGroups

Represents the input of a DescribeCacheSubnetGroups operation.

See: newDescribeCacheSubnetGroups smart constructor.

Constructors

DescribeCacheSubnetGroups'

Fields

cacheSubnetGroupName :: Maybe Text

The name of the cache subnet group to return details for.

maxRecords :: Maybe Int

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: minimum 20; maximum 100.

marker :: Maybe Text

An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

newDescribeCacheSubnetGroups :: DescribeCacheSubnetGroups

Create a value of DescribeCacheSubnetGroups with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cacheSubnetGroupName:DescribeCacheSubnetGroups', describeCacheSubnetGroups_cacheSubnetGroupName - The name of the cache subnet group to return details for.

$sel:maxRecords:DescribeCacheSubnetGroups', describeCacheSubnetGroups_maxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: minimum 20; maximum 100.

$sel:marker:DescribeCacheSubnetGroups', describeCacheSubnetGroups_marker - An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Request Lenses

describeCacheSubnetGroups_cacheSubnetGroupName :: Lens' DescribeCacheSubnetGroups (Maybe Text)

The name of the cache subnet group to return details for.

describeCacheSubnetGroups_maxRecords :: Lens' DescribeCacheSubnetGroups (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: minimum 20; maximum 100.

describeCacheSubnetGroups_marker :: Lens' DescribeCacheSubnetGroups (Maybe Text)

An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Destructuring the Response

data DescribeCacheSubnetGroupsResponse

Represents the output of a DescribeCacheSubnetGroups operation.

See: newDescribeCacheSubnetGroupsResponse smart constructor.

Constructors

DescribeCacheSubnetGroupsResponse'

Fields

cacheSubnetGroups :: Maybe [CacheSubnetGroup]

A list of cache subnet groups. Each element in the list contains detailed information about one group.

marker :: Maybe Text

Provides an identifier to allow retrieval of paginated results.

httpStatus :: Int

The response's http status code.

newDescribeCacheSubnetGroupsResponse

Arguments

:: Int
$sel:httpStatus:DescribeCacheSubnetGroupsResponse'

-> DescribeCacheSubnetGroupsResponse

Create a value of DescribeCacheSubnetGroupsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cacheSubnetGroups:DescribeCacheSubnetGroupsResponse', describeCacheSubnetGroupsResponse_cacheSubnetGroups - A list of cache subnet groups. Each element in the list contains detailed information about one group.

$sel:marker:DescribeCacheSubnetGroups', describeCacheSubnetGroupsResponse_marker - Provides an identifier to allow retrieval of paginated results.

$sel:httpStatus:DescribeCacheSubnetGroupsResponse', describeCacheSubnetGroupsResponse_httpStatus - The response's http status code.

Response Lenses

describeCacheSubnetGroupsResponse_cacheSubnetGroups :: Lens' DescribeCacheSubnetGroupsResponse (Maybe [CacheSubnetGroup])

A list of cache subnet groups. Each element in the list contains detailed information about one group.

describeCacheSubnetGroupsResponse_marker :: Lens' DescribeCacheSubnetGroupsResponse (Maybe Text)

Provides an identifier to allow retrieval of paginated results.

describeCacheSubnetGroupsResponse_httpStatus :: Lens' DescribeCacheSubnetGroupsResponse Int

The response's http status code.