Network.AWS.ElastiCache.DescribeReplicationGroups

Description

Returns information about a particular replication group. If no identifier is specified, DescribeReplicationGroups returns information about all replication groups.

This operation is valid for Redis only.

This operation returns paginated results.

Creating a Request

data DescribeReplicationGroups

Represents the input of a DescribeReplicationGroups operation.

See: newDescribeReplicationGroups smart constructor.

Constructors

DescribeReplicationGroups'

Fields

replicationGroupId :: Maybe Text

The identifier for the replication group to be described. This parameter is not case sensitive.

If you do not specify this parameter, information about all replication groups is returned.

maxRecords :: Maybe Int

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: minimum 20; maximum 100.

marker :: Maybe Text

An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

newDescribeReplicationGroups :: DescribeReplicationGroups

Create a value of DescribeReplicationGroups with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:replicationGroupId:DescribeReplicationGroups', describeReplicationGroups_replicationGroupId - The identifier for the replication group to be described. This parameter is not case sensitive.

If you do not specify this parameter, information about all replication groups is returned.

$sel:maxRecords:DescribeReplicationGroups', describeReplicationGroups_maxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: minimum 20; maximum 100.

$sel:marker:DescribeReplicationGroups', describeReplicationGroups_marker - An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Request Lenses

describeReplicationGroups_replicationGroupId :: Lens' DescribeReplicationGroups (Maybe Text)

The identifier for the replication group to be described. This parameter is not case sensitive.

If you do not specify this parameter, information about all replication groups is returned.

describeReplicationGroups_maxRecords :: Lens' DescribeReplicationGroups (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: minimum 20; maximum 100.

describeReplicationGroups_marker :: Lens' DescribeReplicationGroups (Maybe Text)

An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Destructuring the Response

data DescribeReplicationGroupsResponse

Represents the output of a DescribeReplicationGroups operation.

See: newDescribeReplicationGroupsResponse smart constructor.

Constructors

DescribeReplicationGroupsResponse'

Fields

replicationGroups :: Maybe [ReplicationGroup]

A list of replication groups. Each item in the list contains detailed information about one replication group.

marker :: Maybe Text

Provides an identifier to allow retrieval of paginated results.

httpStatus :: Int

The response's http status code.

newDescribeReplicationGroupsResponse

Arguments

:: Int
$sel:httpStatus:DescribeReplicationGroupsResponse'

-> DescribeReplicationGroupsResponse

Create a value of DescribeReplicationGroupsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:replicationGroups:DescribeReplicationGroupsResponse', describeReplicationGroupsResponse_replicationGroups - A list of replication groups. Each item in the list contains detailed information about one replication group.

$sel:marker:DescribeReplicationGroups', describeReplicationGroupsResponse_marker - Provides an identifier to allow retrieval of paginated results.

$sel:httpStatus:DescribeReplicationGroupsResponse', describeReplicationGroupsResponse_httpStatus - The response's http status code.

Response Lenses

describeReplicationGroupsResponse_replicationGroups :: Lens' DescribeReplicationGroupsResponse (Maybe [ReplicationGroup])

A list of replication groups. Each item in the list contains detailed information about one replication group.

describeReplicationGroupsResponse_marker :: Lens' DescribeReplicationGroupsResponse (Maybe Text)

Provides an identifier to allow retrieval of paginated results.

describeReplicationGroupsResponse_httpStatus :: Lens' DescribeReplicationGroupsResponse Int

The response's http status code.