Network.AWS.ElastiCache.DescribeReservedCacheNodesOfferings

Description

Lists available reserved cache node offerings.

This operation returns paginated results.

Creating a Request

data DescribeReservedCacheNodesOfferings

Represents the input of a DescribeReservedCacheNodesOfferings operation.

See: newDescribeReservedCacheNodesOfferings smart constructor.

Constructors

DescribeReservedCacheNodesOfferings'

Fields

cacheNodeType :: Maybe Text

The cache node type filter value. Use this parameter to show only the available offerings matching the specified cache node type.

The following node types are supported by ElastiCache. Generally speaking, the current generation types provide more memory and computational power at lower cost when compared to their equivalent previous generation counterparts.

General purpose:

Current generation:

M6g node types (available only for Redis engine version 5.0.6 onward and for Memcached engine version 1.5.16 onward).

cache.m6g.large, cache.m6g.xlarge, cache.m6g.2xlarge, cache.m6g.4xlarge, cache.m6g.8xlarge, cache.m6g.12xlarge, cache.m6g.16xlarge

For region availability, see Supported Node Types

M5 node types: cache.m5.large, cache.m5.xlarge, cache.m5.2xlarge, cache.m5.4xlarge, cache.m5.12xlarge, cache.m5.24xlarge

M4 node types: cache.m4.large, cache.m4.xlarge, cache.m4.2xlarge, cache.m4.4xlarge, cache.m4.10xlarge

T3 node types: cache.t3.micro, cache.t3.small, cache.t3.medium

T2 node types: cache.t2.micro, cache.t2.small, cache.t2.medium

Previous generation: (not recommended)

T1 node types: cache.t1.micro

M1 node types: cache.m1.small, cache.m1.medium, cache.m1.large, cache.m1.xlarge

M3 node types: cache.m3.medium, cache.m3.large, cache.m3.xlarge, cache.m3.2xlarge

Compute optimized:

Previous generation: (not recommended)

C1 node types: cache.c1.xlarge

Memory optimized:

Current generation:

R6g node types (available only for Redis engine version 5.0.6 onward and for Memcached engine version 1.5.16 onward).

cache.r6g.large, cache.r6g.xlarge, cache.r6g.2xlarge, cache.r6g.4xlarge, cache.r6g.8xlarge, cache.r6g.12xlarge, cache.r6g.16xlarge

For region availability, see Supported Node Types

R5 node types: cache.r5.large, cache.r5.xlarge, cache.r5.2xlarge, cache.r5.4xlarge, cache.r5.12xlarge, cache.r5.24xlarge

R4 node types: cache.r4.large, cache.r4.xlarge, cache.r4.2xlarge, cache.r4.4xlarge, cache.r4.8xlarge, cache.r4.16xlarge

Previous generation: (not recommended)

M2 node types: cache.m2.xlarge, cache.m2.2xlarge, cache.m2.4xlarge

R3 node types: cache.r3.large, cache.r3.xlarge, cache.r3.2xlarge, cache.r3.4xlarge, cache.r3.8xlarge

Additional node type info

All current generation instance types are created in Amazon VPC by default.

Redis append-only files (AOF) are not supported for T1 or T2 instances.

Redis Multi-AZ with automatic failover is not supported on T1 instances.

Redis configuration variables appendonly and appendfsync are not supported on Redis version 2.8.22 and later.

productDescription :: Maybe Text

The product description filter value. Use this parameter to show only the available offerings matching the specified product description.

marker :: Maybe Text

An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

maxRecords :: Maybe Int

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: minimum 20; maximum 100.

offeringType :: Maybe Text

The offering type filter value. Use this parameter to show only the available offerings matching the specified offering type.

Valid Values: "Light Utilization"|"Medium Utilization"|"Heavy Utilization" |"All Upfront"|"Partial Upfront"| "No Upfront"

duration :: Maybe Text

Duration filter value, specified in years or seconds. Use this parameter to show only reservations for a given duration.

Valid Values: 1 | 3 | 31536000 | 94608000

reservedCacheNodesOfferingId :: Maybe Text

The offering identifier filter value. Use this parameter to show only the available offering that matches the specified reservation identifier.

Example: 438012d3-4052-4cc7-b2e3-8d3372e0e706

newDescribeReservedCacheNodesOfferings :: DescribeReservedCacheNodesOfferings

Create a value of DescribeReservedCacheNodesOfferings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cacheNodeType:DescribeReservedCacheNodesOfferings', describeReservedCacheNodesOfferings_cacheNodeType - The cache node type filter value. Use this parameter to show only the available offerings matching the specified cache node type.

The following node types are supported by ElastiCache. Generally speaking, the current generation types provide more memory and computational power at lower cost when compared to their equivalent previous generation counterparts.

General purpose:

Current generation:

M6g node types (available only for Redis engine version 5.0.6 onward and for Memcached engine version 1.5.16 onward).

cache.m6g.large, cache.m6g.xlarge, cache.m6g.2xlarge, cache.m6g.4xlarge, cache.m6g.8xlarge, cache.m6g.12xlarge, cache.m6g.16xlarge

For region availability, see Supported Node Types

M5 node types: cache.m5.large, cache.m5.xlarge, cache.m5.2xlarge, cache.m5.4xlarge, cache.m5.12xlarge, cache.m5.24xlarge

M4 node types: cache.m4.large, cache.m4.xlarge, cache.m4.2xlarge, cache.m4.4xlarge, cache.m4.10xlarge

T3 node types: cache.t3.micro, cache.t3.small, cache.t3.medium

T2 node types: cache.t2.micro, cache.t2.small, cache.t2.medium

Previous generation: (not recommended)

T1 node types: cache.t1.micro

M1 node types: cache.m1.small, cache.m1.medium, cache.m1.large, cache.m1.xlarge

M3 node types: cache.m3.medium, cache.m3.large, cache.m3.xlarge, cache.m3.2xlarge

Compute optimized:

Previous generation: (not recommended)

C1 node types: cache.c1.xlarge

Memory optimized:

Current generation:

R6g node types (available only for Redis engine version 5.0.6 onward and for Memcached engine version 1.5.16 onward).

cache.r6g.large, cache.r6g.xlarge, cache.r6g.2xlarge, cache.r6g.4xlarge, cache.r6g.8xlarge, cache.r6g.12xlarge, cache.r6g.16xlarge

For region availability, see Supported Node Types

R5 node types: cache.r5.large, cache.r5.xlarge, cache.r5.2xlarge, cache.r5.4xlarge, cache.r5.12xlarge, cache.r5.24xlarge

R4 node types: cache.r4.large, cache.r4.xlarge, cache.r4.2xlarge, cache.r4.4xlarge, cache.r4.8xlarge, cache.r4.16xlarge

Previous generation: (not recommended)

M2 node types: cache.m2.xlarge, cache.m2.2xlarge, cache.m2.4xlarge

R3 node types: cache.r3.large, cache.r3.xlarge, cache.r3.2xlarge, cache.r3.4xlarge, cache.r3.8xlarge

Additional node type info

All current generation instance types are created in Amazon VPC by default.

Redis append-only files (AOF) are not supported for T1 or T2 instances.

Redis Multi-AZ with automatic failover is not supported on T1 instances.

Redis configuration variables appendonly and appendfsync are not supported on Redis version 2.8.22 and later.

$sel:productDescription:DescribeReservedCacheNodesOfferings', describeReservedCacheNodesOfferings_productDescription - The product description filter value. Use this parameter to show only the available offerings matching the specified product description.

$sel:marker:DescribeReservedCacheNodesOfferings', describeReservedCacheNodesOfferings_marker - An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

$sel:maxRecords:DescribeReservedCacheNodesOfferings', describeReservedCacheNodesOfferings_maxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: minimum 20; maximum 100.

$sel:offeringType:DescribeReservedCacheNodesOfferings', describeReservedCacheNodesOfferings_offeringType - The offering type filter value. Use this parameter to show only the available offerings matching the specified offering type.

Valid Values: "Light Utilization"|"Medium Utilization"|"Heavy Utilization" |"All Upfront"|"Partial Upfront"| "No Upfront"

$sel:duration:DescribeReservedCacheNodesOfferings', describeReservedCacheNodesOfferings_duration - Duration filter value, specified in years or seconds. Use this parameter to show only reservations for a given duration.

Valid Values: 1 | 3 | 31536000 | 94608000

$sel:reservedCacheNodesOfferingId:DescribeReservedCacheNodesOfferings', describeReservedCacheNodesOfferings_reservedCacheNodesOfferingId - The offering identifier filter value. Use this parameter to show only the available offering that matches the specified reservation identifier.

Example: 438012d3-4052-4cc7-b2e3-8d3372e0e706

Request Lenses

describeReservedCacheNodesOfferings_cacheNodeType :: Lens' DescribeReservedCacheNodesOfferings (Maybe Text)

The cache node type filter value. Use this parameter to show only the available offerings matching the specified cache node type.

The following node types are supported by ElastiCache. Generally speaking, the current generation types provide more memory and computational power at lower cost when compared to their equivalent previous generation counterparts.

General purpose:

Current generation:

M6g node types (available only for Redis engine version 5.0.6 onward and for Memcached engine version 1.5.16 onward).

cache.m6g.large, cache.m6g.xlarge, cache.m6g.2xlarge, cache.m6g.4xlarge, cache.m6g.8xlarge, cache.m6g.12xlarge, cache.m6g.16xlarge

For region availability, see Supported Node Types

M5 node types: cache.m5.large, cache.m5.xlarge, cache.m5.2xlarge, cache.m5.4xlarge, cache.m5.12xlarge, cache.m5.24xlarge

M4 node types: cache.m4.large, cache.m4.xlarge, cache.m4.2xlarge, cache.m4.4xlarge, cache.m4.10xlarge

T3 node types: cache.t3.micro, cache.t3.small, cache.t3.medium

T2 node types: cache.t2.micro, cache.t2.small, cache.t2.medium

Previous generation: (not recommended)

T1 node types: cache.t1.micro

M1 node types: cache.m1.small, cache.m1.medium, cache.m1.large, cache.m1.xlarge

M3 node types: cache.m3.medium, cache.m3.large, cache.m3.xlarge, cache.m3.2xlarge

Compute optimized:

Previous generation: (not recommended)

C1 node types: cache.c1.xlarge

Memory optimized:

Current generation:

R6g node types (available only for Redis engine version 5.0.6 onward and for Memcached engine version 1.5.16 onward).

cache.r6g.large, cache.r6g.xlarge, cache.r6g.2xlarge, cache.r6g.4xlarge, cache.r6g.8xlarge, cache.r6g.12xlarge, cache.r6g.16xlarge

For region availability, see Supported Node Types

R5 node types: cache.r5.large, cache.r5.xlarge, cache.r5.2xlarge, cache.r5.4xlarge, cache.r5.12xlarge, cache.r5.24xlarge

R4 node types: cache.r4.large, cache.r4.xlarge, cache.r4.2xlarge, cache.r4.4xlarge, cache.r4.8xlarge, cache.r4.16xlarge

Previous generation: (not recommended)

M2 node types: cache.m2.xlarge, cache.m2.2xlarge, cache.m2.4xlarge

R3 node types: cache.r3.large, cache.r3.xlarge, cache.r3.2xlarge, cache.r3.4xlarge, cache.r3.8xlarge

Additional node type info

All current generation instance types are created in Amazon VPC by default.

Redis append-only files (AOF) are not supported for T1 or T2 instances.

Redis Multi-AZ with automatic failover is not supported on T1 instances.

Redis configuration variables appendonly and appendfsync are not supported on Redis version 2.8.22 and later.

describeReservedCacheNodesOfferings_productDescription :: Lens' DescribeReservedCacheNodesOfferings (Maybe Text)

The product description filter value. Use this parameter to show only the available offerings matching the specified product description.

describeReservedCacheNodesOfferings_marker :: Lens' DescribeReservedCacheNodesOfferings (Maybe Text)

An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

describeReservedCacheNodesOfferings_maxRecords :: Lens' DescribeReservedCacheNodesOfferings (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: minimum 20; maximum 100.

describeReservedCacheNodesOfferings_offeringType :: Lens' DescribeReservedCacheNodesOfferings (Maybe Text)

The offering type filter value. Use this parameter to show only the available offerings matching the specified offering type.

Valid Values: "Light Utilization"|"Medium Utilization"|"Heavy Utilization" |"All Upfront"|"Partial Upfront"| "No Upfront"

describeReservedCacheNodesOfferings_duration :: Lens' DescribeReservedCacheNodesOfferings (Maybe Text)

Duration filter value, specified in years or seconds. Use this parameter to show only reservations for a given duration.

Valid Values: 1 | 3 | 31536000 | 94608000

describeReservedCacheNodesOfferings_reservedCacheNodesOfferingId :: Lens' DescribeReservedCacheNodesOfferings (Maybe Text)

The offering identifier filter value. Use this parameter to show only the available offering that matches the specified reservation identifier.

Example: 438012d3-4052-4cc7-b2e3-8d3372e0e706

Destructuring the Response

data DescribeReservedCacheNodesOfferingsResponse

Represents the output of a DescribeReservedCacheNodesOfferings operation.

See: newDescribeReservedCacheNodesOfferingsResponse smart constructor.

Constructors

DescribeReservedCacheNodesOfferingsResponse'

Fields

marker :: Maybe Text

Provides an identifier to allow retrieval of paginated results.

reservedCacheNodesOfferings :: Maybe [ReservedCacheNodesOffering]

A list of reserved cache node offerings. Each element in the list contains detailed information about one offering.

httpStatus :: Int

The response's http status code.

newDescribeReservedCacheNodesOfferingsResponse

Arguments

:: Int
$sel:httpStatus:DescribeReservedCacheNodesOfferingsResponse'

-> DescribeReservedCacheNodesOfferingsResponse

Create a value of DescribeReservedCacheNodesOfferingsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:marker:DescribeReservedCacheNodesOfferings', describeReservedCacheNodesOfferingsResponse_marker - Provides an identifier to allow retrieval of paginated results.

$sel:reservedCacheNodesOfferings:DescribeReservedCacheNodesOfferingsResponse', describeReservedCacheNodesOfferingsResponse_reservedCacheNodesOfferings - A list of reserved cache node offerings. Each element in the list contains detailed information about one offering.

$sel:httpStatus:DescribeReservedCacheNodesOfferingsResponse', describeReservedCacheNodesOfferingsResponse_httpStatus - The response's http status code.

Response Lenses

describeReservedCacheNodesOfferingsResponse_marker :: Lens' DescribeReservedCacheNodesOfferingsResponse (Maybe Text)

Provides an identifier to allow retrieval of paginated results.

describeReservedCacheNodesOfferingsResponse_reservedCacheNodesOfferings :: Lens' DescribeReservedCacheNodesOfferingsResponse (Maybe [ReservedCacheNodesOffering])

A list of reserved cache node offerings. Each element in the list contains detailed information about one offering.

describeReservedCacheNodesOfferingsResponse_httpStatus :: Lens' DescribeReservedCacheNodesOfferingsResponse Int

The response's http status code.