Network.AWS.ElastiCache.DescribeServiceUpdates

Description

Returns details of the service updates

This operation returns paginated results.

Creating a Request

data DescribeServiceUpdates

See: newDescribeServiceUpdates smart constructor.

Constructors

DescribeServiceUpdates'

Fields

serviceUpdateStatus :: Maybe [ServiceUpdateStatus]

The status of the service update

serviceUpdateName :: Maybe Text

The unique ID of the service update

maxRecords :: Maybe Int

The maximum number of records to include in the response

marker :: Maybe Text

An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

newDescribeServiceUpdates :: DescribeServiceUpdates

Create a value of DescribeServiceUpdates with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceUpdateStatus:DescribeServiceUpdates', describeServiceUpdates_serviceUpdateStatus - The status of the service update

$sel:serviceUpdateName:DescribeServiceUpdates', describeServiceUpdates_serviceUpdateName - The unique ID of the service update

$sel:maxRecords:DescribeServiceUpdates', describeServiceUpdates_maxRecords - The maximum number of records to include in the response

$sel:marker:DescribeServiceUpdates', describeServiceUpdates_marker - An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Request Lenses

describeServiceUpdates_serviceUpdateStatus :: Lens' DescribeServiceUpdates (Maybe [ServiceUpdateStatus])

The status of the service update

describeServiceUpdates_serviceUpdateName :: Lens' DescribeServiceUpdates (Maybe Text)

The unique ID of the service update

describeServiceUpdates_maxRecords :: Lens' DescribeServiceUpdates (Maybe Int)

The maximum number of records to include in the response

describeServiceUpdates_marker :: Lens' DescribeServiceUpdates (Maybe Text)

An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Destructuring the Response

data DescribeServiceUpdatesResponse

See: newDescribeServiceUpdatesResponse smart constructor.

Constructors

DescribeServiceUpdatesResponse'

Fields

serviceUpdates :: Maybe [ServiceUpdate]

A list of service updates

marker :: Maybe Text

An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

httpStatus :: Int

The response's http status code.

newDescribeServiceUpdatesResponse

Arguments

:: Int
$sel:httpStatus:DescribeServiceUpdatesResponse'

-> DescribeServiceUpdatesResponse

Create a value of DescribeServiceUpdatesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceUpdates:DescribeServiceUpdatesResponse', describeServiceUpdatesResponse_serviceUpdates - A list of service updates

$sel:marker:DescribeServiceUpdates', describeServiceUpdatesResponse_marker - An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

$sel:httpStatus:DescribeServiceUpdatesResponse', describeServiceUpdatesResponse_httpStatus - The response's http status code.

Response Lenses

describeServiceUpdatesResponse_serviceUpdates :: Lens' DescribeServiceUpdatesResponse (Maybe [ServiceUpdate])

A list of service updates

describeServiceUpdatesResponse_marker :: Lens' DescribeServiceUpdatesResponse (Maybe Text)

An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

describeServiceUpdatesResponse_httpStatus :: Lens' DescribeServiceUpdatesResponse Int

The response's http status code.