Network.AWS.ElastiCache.DescribeUpdateActions

Description

Returns details of the update actions

This operation returns paginated results.

Creating a Request

data DescribeUpdateActions

See: newDescribeUpdateActions smart constructor.

Constructors

DescribeUpdateActions'

Fields

updateActionStatus :: Maybe [UpdateActionStatus]

The status of the update action.

serviceUpdateStatus :: Maybe [ServiceUpdateStatus]

The status of the service update

showNodeLevelUpdateStatus :: Maybe Bool

Dictates whether to include node level update status in the response

engine :: Maybe Text

The Elasticache engine to which the update applies. Either Redis or Memcached

serviceUpdateTimeRange :: Maybe TimeRangeFilter

The range of time specified to search for service updates that are in available status

serviceUpdateName :: Maybe Text

The unique ID of the service update

cacheClusterIds :: Maybe [Text]

The cache cluster IDs

replicationGroupIds :: Maybe [Text]

The replication group IDs

maxRecords :: Maybe Int

The maximum number of records to include in the response

marker :: Maybe Text

An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

newDescribeUpdateActions :: DescribeUpdateActions

Create a value of DescribeUpdateActions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:updateActionStatus:DescribeUpdateActions', describeUpdateActions_updateActionStatus - The status of the update action.

$sel:serviceUpdateStatus:DescribeUpdateActions', describeUpdateActions_serviceUpdateStatus - The status of the service update

$sel:showNodeLevelUpdateStatus:DescribeUpdateActions', describeUpdateActions_showNodeLevelUpdateStatus - Dictates whether to include node level update status in the response

$sel:engine:DescribeUpdateActions', describeUpdateActions_engine - The Elasticache engine to which the update applies. Either Redis or Memcached

$sel:serviceUpdateTimeRange:DescribeUpdateActions', describeUpdateActions_serviceUpdateTimeRange - The range of time specified to search for service updates that are in available status

$sel:serviceUpdateName:DescribeUpdateActions', describeUpdateActions_serviceUpdateName - The unique ID of the service update

$sel:cacheClusterIds:DescribeUpdateActions', describeUpdateActions_cacheClusterIds - The cache cluster IDs

$sel:replicationGroupIds:DescribeUpdateActions', describeUpdateActions_replicationGroupIds - The replication group IDs

$sel:maxRecords:DescribeUpdateActions', describeUpdateActions_maxRecords - The maximum number of records to include in the response

$sel:marker:DescribeUpdateActions', describeUpdateActions_marker - An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Request Lenses

describeUpdateActions_updateActionStatus :: Lens' DescribeUpdateActions (Maybe [UpdateActionStatus])

The status of the update action.

describeUpdateActions_serviceUpdateStatus :: Lens' DescribeUpdateActions (Maybe [ServiceUpdateStatus])

The status of the service update

describeUpdateActions_showNodeLevelUpdateStatus :: Lens' DescribeUpdateActions (Maybe Bool)

Dictates whether to include node level update status in the response

describeUpdateActions_engine :: Lens' DescribeUpdateActions (Maybe Text)

The Elasticache engine to which the update applies. Either Redis or Memcached

describeUpdateActions_serviceUpdateTimeRange :: Lens' DescribeUpdateActions (Maybe TimeRangeFilter)

The range of time specified to search for service updates that are in available status

describeUpdateActions_serviceUpdateName :: Lens' DescribeUpdateActions (Maybe Text)

The unique ID of the service update

describeUpdateActions_cacheClusterIds :: Lens' DescribeUpdateActions (Maybe [Text])

The cache cluster IDs

describeUpdateActions_replicationGroupIds :: Lens' DescribeUpdateActions (Maybe [Text])

The replication group IDs

describeUpdateActions_maxRecords :: Lens' DescribeUpdateActions (Maybe Int)

The maximum number of records to include in the response

describeUpdateActions_marker :: Lens' DescribeUpdateActions (Maybe Text)

An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Destructuring the Response

data DescribeUpdateActionsResponse

See: newDescribeUpdateActionsResponse smart constructor.

Constructors

DescribeUpdateActionsResponse'

Fields

updateActions :: Maybe [UpdateAction]

Returns a list of update actions

marker :: Maybe Text

An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

httpStatus :: Int

The response's http status code.

newDescribeUpdateActionsResponse

Arguments

:: Int
$sel:httpStatus:DescribeUpdateActionsResponse'

-> DescribeUpdateActionsResponse

Create a value of DescribeUpdateActionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:updateActions:DescribeUpdateActionsResponse', describeUpdateActionsResponse_updateActions - Returns a list of update actions

$sel:marker:DescribeUpdateActions', describeUpdateActionsResponse_marker - An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

$sel:httpStatus:DescribeUpdateActionsResponse', describeUpdateActionsResponse_httpStatus - The response's http status code.

Response Lenses

describeUpdateActionsResponse_updateActions :: Lens' DescribeUpdateActionsResponse (Maybe [UpdateAction])

Returns a list of update actions

describeUpdateActionsResponse_marker :: Lens' DescribeUpdateActionsResponse (Maybe Text)

An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

describeUpdateActionsResponse_httpStatus :: Lens' DescribeUpdateActionsResponse Int

The response's http status code.