Network.AWS.ElastiCache.DescribeUserGroups

Description

Returns a list of user groups.

This operation returns paginated results.

Creating a Request

data DescribeUserGroups

See: newDescribeUserGroups smart constructor.

Constructors

DescribeUserGroups'

Fields

userGroupId :: Maybe Text

The ID of the user group.

maxRecords :: Maybe Int

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved.

marker :: Maybe Text

An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords. >

newDescribeUserGroups :: DescribeUserGroups

Create a value of DescribeUserGroups with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userGroupId:DescribeUserGroups', describeUserGroups_userGroupId - The ID of the user group.

$sel:maxRecords:DescribeUserGroups', describeUserGroups_maxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved.

$sel:marker:DescribeUserGroups', describeUserGroups_marker - An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords. >

Request Lenses

describeUserGroups_userGroupId :: Lens' DescribeUserGroups (Maybe Text)

The ID of the user group.

describeUserGroups_maxRecords :: Lens' DescribeUserGroups (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved.

describeUserGroups_marker :: Lens' DescribeUserGroups (Maybe Text)

An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords. >

Destructuring the Response

data DescribeUserGroupsResponse

See: newDescribeUserGroupsResponse smart constructor.

Constructors

DescribeUserGroupsResponse'

Fields

userGroups :: Maybe [UserGroup]

Returns a list of user groups.

marker :: Maybe Text

An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords. >

httpStatus :: Int

The response's http status code.

newDescribeUserGroupsResponse

Arguments

:: Int
$sel:httpStatus:DescribeUserGroupsResponse'

-> DescribeUserGroupsResponse

Create a value of DescribeUserGroupsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userGroups:DescribeUserGroupsResponse', describeUserGroupsResponse_userGroups - Returns a list of user groups.

$sel:marker:DescribeUserGroups', describeUserGroupsResponse_marker - An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords. >

$sel:httpStatus:DescribeUserGroupsResponse', describeUserGroupsResponse_httpStatus - The response's http status code.

Response Lenses

describeUserGroupsResponse_userGroups :: Lens' DescribeUserGroupsResponse (Maybe [UserGroup])

Returns a list of user groups.

describeUserGroupsResponse_marker :: Lens' DescribeUserGroupsResponse (Maybe Text)

An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords. >

describeUserGroupsResponse_httpStatus :: Lens' DescribeUserGroupsResponse Int

The response's http status code.