Network.AWS.ElastiCache.DisassociateGlobalReplicationGroup

Description

Remove a secondary cluster from the Global datastore using the Global datastore name. The secondary cluster will no longer receive updates from the primary cluster, but will remain as a standalone cluster in that Amazon region.

Creating a Request

data DisassociateGlobalReplicationGroup

See: newDisassociateGlobalReplicationGroup smart constructor.

Constructors

DisassociateGlobalReplicationGroup'

Fields

globalReplicationGroupId :: Text

The name of the Global datastore

replicationGroupId :: Text

The name of the secondary cluster you wish to remove from the Global datastore

replicationGroupRegion :: Text

The Amazon region of secondary cluster you wish to remove from the Global datastore

newDisassociateGlobalReplicationGroup

Arguments

:: Text
$sel:globalReplicationGroupId:DisassociateGlobalReplicationGroup'

-> Text
$sel:replicationGroupId:DisassociateGlobalReplicationGroup'

-> Text
$sel:replicationGroupRegion:DisassociateGlobalReplicationGroup'

-> DisassociateGlobalReplicationGroup

Create a value of DisassociateGlobalReplicationGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:globalReplicationGroupId:DisassociateGlobalReplicationGroup', disassociateGlobalReplicationGroup_globalReplicationGroupId - The name of the Global datastore

$sel:replicationGroupId:DisassociateGlobalReplicationGroup', disassociateGlobalReplicationGroup_replicationGroupId - The name of the secondary cluster you wish to remove from the Global datastore

$sel:replicationGroupRegion:DisassociateGlobalReplicationGroup', disassociateGlobalReplicationGroup_replicationGroupRegion - The Amazon region of secondary cluster you wish to remove from the Global datastore

Request Lenses

disassociateGlobalReplicationGroup_globalReplicationGroupId :: Lens' DisassociateGlobalReplicationGroup Text

The name of the Global datastore

disassociateGlobalReplicationGroup_replicationGroupId :: Lens' DisassociateGlobalReplicationGroup Text

The name of the secondary cluster you wish to remove from the Global datastore

disassociateGlobalReplicationGroup_replicationGroupRegion :: Lens' DisassociateGlobalReplicationGroup Text

The Amazon region of secondary cluster you wish to remove from the Global datastore

Destructuring the Response

data DisassociateGlobalReplicationGroupResponse

See: newDisassociateGlobalReplicationGroupResponse smart constructor.

Constructors

DisassociateGlobalReplicationGroupResponse'

Fields

globalReplicationGroup :: Maybe GlobalReplicationGroup

httpStatus :: Int

The response's http status code.

newDisassociateGlobalReplicationGroupResponse

Arguments

:: Int
$sel:httpStatus:DisassociateGlobalReplicationGroupResponse'

-> DisassociateGlobalReplicationGroupResponse

Create a value of DisassociateGlobalReplicationGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:globalReplicationGroup:DisassociateGlobalReplicationGroupResponse', disassociateGlobalReplicationGroupResponse_globalReplicationGroup - Undocumented member.

$sel:httpStatus:DisassociateGlobalReplicationGroupResponse', disassociateGlobalReplicationGroupResponse_httpStatus - The response's http status code.

Response Lenses

disassociateGlobalReplicationGroupResponse_globalReplicationGroup :: Lens' DisassociateGlobalReplicationGroupResponse (Maybe GlobalReplicationGroup)

Undocumented member.

disassociateGlobalReplicationGroupResponse_httpStatus :: Lens' DisassociateGlobalReplicationGroupResponse Int

The response's http status code.