Network.AWS.ElastiCache.IncreaseNodeGroupsInGlobalReplicationGroup

Description

Increase the number of node groups in the Global datastore

Creating a Request

data IncreaseNodeGroupsInGlobalReplicationGroup

See: newIncreaseNodeGroupsInGlobalReplicationGroup smart constructor.

Constructors

IncreaseNodeGroupsInGlobalReplicationGroup'

Fields

regionalConfigurations :: Maybe [RegionalConfiguration]

Describes the replication group IDs, the Amazon regions where they are stored and the shard configuration for each that comprise the Global datastore

globalReplicationGroupId :: Text

The name of the Global datastore

nodeGroupCount :: Int

The number of node groups you wish to add

applyImmediately :: Bool

Indicates that the process begins immediately. At present, the only permitted value for this parameter is true.

newIncreaseNodeGroupsInGlobalReplicationGroup

Arguments

:: Text
$sel:globalReplicationGroupId:IncreaseNodeGroupsInGlobalReplicationGroup'

-> Int
$sel:nodeGroupCount:IncreaseNodeGroupsInGlobalReplicationGroup'

-> Bool
$sel:applyImmediately:IncreaseNodeGroupsInGlobalReplicationGroup'

-> IncreaseNodeGroupsInGlobalReplicationGroup

Create a value of IncreaseNodeGroupsInGlobalReplicationGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:regionalConfigurations:IncreaseNodeGroupsInGlobalReplicationGroup', increaseNodeGroupsInGlobalReplicationGroup_regionalConfigurations - Describes the replication group IDs, the Amazon regions where they are stored and the shard configuration for each that comprise the Global datastore

$sel:globalReplicationGroupId:IncreaseNodeGroupsInGlobalReplicationGroup', increaseNodeGroupsInGlobalReplicationGroup_globalReplicationGroupId - The name of the Global datastore

$sel:nodeGroupCount:IncreaseNodeGroupsInGlobalReplicationGroup', increaseNodeGroupsInGlobalReplicationGroup_nodeGroupCount - The number of node groups you wish to add

$sel:applyImmediately:IncreaseNodeGroupsInGlobalReplicationGroup', increaseNodeGroupsInGlobalReplicationGroup_applyImmediately - Indicates that the process begins immediately. At present, the only permitted value for this parameter is true.

Request Lenses

increaseNodeGroupsInGlobalReplicationGroup_regionalConfigurations :: Lens' IncreaseNodeGroupsInGlobalReplicationGroup (Maybe [RegionalConfiguration])

Describes the replication group IDs, the Amazon regions where they are stored and the shard configuration for each that comprise the Global datastore

increaseNodeGroupsInGlobalReplicationGroup_globalReplicationGroupId :: Lens' IncreaseNodeGroupsInGlobalReplicationGroup Text

The name of the Global datastore

increaseNodeGroupsInGlobalReplicationGroup_nodeGroupCount :: Lens' IncreaseNodeGroupsInGlobalReplicationGroup Int

The number of node groups you wish to add

increaseNodeGroupsInGlobalReplicationGroup_applyImmediately :: Lens' IncreaseNodeGroupsInGlobalReplicationGroup Bool

Indicates that the process begins immediately. At present, the only permitted value for this parameter is true.

Destructuring the Response

data IncreaseNodeGroupsInGlobalReplicationGroupResponse

See: newIncreaseNodeGroupsInGlobalReplicationGroupResponse smart constructor.

Constructors

IncreaseNodeGroupsInGlobalReplicationGroupResponse'

Fields

globalReplicationGroup :: Maybe GlobalReplicationGroup

httpStatus :: Int

The response's http status code.

newIncreaseNodeGroupsInGlobalReplicationGroupResponse

Arguments

:: Int
$sel:httpStatus:IncreaseNodeGroupsInGlobalReplicationGroupResponse'

-> IncreaseNodeGroupsInGlobalReplicationGroupResponse

Create a value of IncreaseNodeGroupsInGlobalReplicationGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:globalReplicationGroup:IncreaseNodeGroupsInGlobalReplicationGroupResponse', increaseNodeGroupsInGlobalReplicationGroupResponse_globalReplicationGroup - Undocumented member.

$sel:httpStatus:IncreaseNodeGroupsInGlobalReplicationGroupResponse', increaseNodeGroupsInGlobalReplicationGroupResponse_httpStatus - The response's http status code.

Response Lenses

increaseNodeGroupsInGlobalReplicationGroupResponse_globalReplicationGroup :: Lens' IncreaseNodeGroupsInGlobalReplicationGroupResponse (Maybe GlobalReplicationGroup)

Undocumented member.

increaseNodeGroupsInGlobalReplicationGroupResponse_httpStatus :: Lens' IncreaseNodeGroupsInGlobalReplicationGroupResponse Int

The response's http status code.