Network.AWS.ElastiCache.ListAllowedNodeTypeModifications

Description

Lists all available node types that you can scale your Redis cluster's or replication group's current node type.

When you use the ModifyCacheCluster or ModifyReplicationGroup operations to scale your cluster or replication group, the value of the CacheNodeType parameter must be one of the node types returned by this operation.

Creating a Request

data ListAllowedNodeTypeModifications

The input parameters for the ListAllowedNodeTypeModifications operation.

See: newListAllowedNodeTypeModifications smart constructor.

Constructors

ListAllowedNodeTypeModifications'

Fields

replicationGroupId :: Maybe Text

The name of the replication group want to scale up to a larger node type. ElastiCache uses the replication group id to identify the current node type being used by this replication group, and from that to create a list of node types you can scale up to.

You must provide a value for either the CacheClusterId or the ReplicationGroupId.

cacheClusterId :: Maybe Text

The name of the cluster you want to scale up to a larger node instanced type. ElastiCache uses the cluster id to identify the current node type of this cluster and from that to create a list of node types you can scale up to.

You must provide a value for either the CacheClusterId or the ReplicationGroupId.

newListAllowedNodeTypeModifications :: ListAllowedNodeTypeModifications

Create a value of ListAllowedNodeTypeModifications with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:replicationGroupId:ListAllowedNodeTypeModifications', listAllowedNodeTypeModifications_replicationGroupId - The name of the replication group want to scale up to a larger node type. ElastiCache uses the replication group id to identify the current node type being used by this replication group, and from that to create a list of node types you can scale up to.

You must provide a value for either the CacheClusterId or the ReplicationGroupId.

$sel:cacheClusterId:ListAllowedNodeTypeModifications', listAllowedNodeTypeModifications_cacheClusterId - The name of the cluster you want to scale up to a larger node instanced type. ElastiCache uses the cluster id to identify the current node type of this cluster and from that to create a list of node types you can scale up to.

You must provide a value for either the CacheClusterId or the ReplicationGroupId.

Request Lenses

listAllowedNodeTypeModifications_replicationGroupId :: Lens' ListAllowedNodeTypeModifications (Maybe Text)

The name of the replication group want to scale up to a larger node type. ElastiCache uses the replication group id to identify the current node type being used by this replication group, and from that to create a list of node types you can scale up to.

You must provide a value for either the CacheClusterId or the ReplicationGroupId.

listAllowedNodeTypeModifications_cacheClusterId :: Lens' ListAllowedNodeTypeModifications (Maybe Text)

The name of the cluster you want to scale up to a larger node instanced type. ElastiCache uses the cluster id to identify the current node type of this cluster and from that to create a list of node types you can scale up to.

You must provide a value for either the CacheClusterId or the ReplicationGroupId.

Destructuring the Response

data ListAllowedNodeTypeModificationsResponse

Represents the allowed node types you can use to modify your cluster or replication group.

See: newListAllowedNodeTypeModificationsResponse smart constructor.

Constructors

ListAllowedNodeTypeModificationsResponse'

Fields

scaleUpModifications :: Maybe [Text]

A string list, each element of which specifies a cache node type which you can use to scale your cluster or replication group.

When scaling up a Redis cluster or replication group using ModifyCacheCluster or ModifyReplicationGroup, use a value from this list for the CacheNodeType parameter.

scaleDownModifications :: Maybe [Text]

A string list, each element of which specifies a cache node type which you can use to scale your cluster or replication group. When scaling down a Redis cluster or replication group using ModifyCacheCluster or ModifyReplicationGroup, use a value from this list for the CacheNodeType parameter.

httpStatus :: Int

The response's http status code.

newListAllowedNodeTypeModificationsResponse

Arguments

:: Int
$sel:httpStatus:ListAllowedNodeTypeModificationsResponse'

-> ListAllowedNodeTypeModificationsResponse

Create a value of ListAllowedNodeTypeModificationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:scaleUpModifications:ListAllowedNodeTypeModificationsResponse', listAllowedNodeTypeModificationsResponse_scaleUpModifications - A string list, each element of which specifies a cache node type which you can use to scale your cluster or replication group.

When scaling up a Redis cluster or replication group using ModifyCacheCluster or ModifyReplicationGroup, use a value from this list for the CacheNodeType parameter.

$sel:scaleDownModifications:ListAllowedNodeTypeModificationsResponse', listAllowedNodeTypeModificationsResponse_scaleDownModifications - A string list, each element of which specifies a cache node type which you can use to scale your cluster or replication group. When scaling down a Redis cluster or replication group using ModifyCacheCluster or ModifyReplicationGroup, use a value from this list for the CacheNodeType parameter.

$sel:httpStatus:ListAllowedNodeTypeModificationsResponse', listAllowedNodeTypeModificationsResponse_httpStatus - The response's http status code.

Response Lenses

listAllowedNodeTypeModificationsResponse_scaleUpModifications :: Lens' ListAllowedNodeTypeModificationsResponse (Maybe [Text])

A string list, each element of which specifies a cache node type which you can use to scale your cluster or replication group.

When scaling up a Redis cluster or replication group using ModifyCacheCluster or ModifyReplicationGroup, use a value from this list for the CacheNodeType parameter.

listAllowedNodeTypeModificationsResponse_scaleDownModifications :: Lens' ListAllowedNodeTypeModificationsResponse (Maybe [Text])

A string list, each element of which specifies a cache node type which you can use to scale your cluster or replication group. When scaling down a Redis cluster or replication group using ModifyCacheCluster or ModifyReplicationGroup, use a value from this list for the CacheNodeType parameter.

listAllowedNodeTypeModificationsResponse_httpStatus :: Lens' ListAllowedNodeTypeModificationsResponse Int

The response's http status code.