Network.AWS.ElastiCache.ModifyGlobalReplicationGroup

Description

Modifies the settings for a Global datastore.

Creating a Request

data ModifyGlobalReplicationGroup

See: newModifyGlobalReplicationGroup smart constructor.

Constructors

ModifyGlobalReplicationGroup'

Fields

automaticFailoverEnabled :: Maybe Bool

Determines whether a read replica is automatically promoted to read/write primary if the existing primary encounters a failure.

cacheParameterGroupName :: Maybe Text

The name of the cache parameter group to use with the Global datastore. It must be compatible with the major engine version used by the Global datastore.

engineVersion :: Maybe Text

The upgraded version of the cache engine to be run on the clusters in the Global datastore.

cacheNodeType :: Maybe Text

A valid cache node type that you want to scale this Global datastore to.

globalReplicationGroupDescription :: Maybe Text

A description of the Global datastore

globalReplicationGroupId :: Text

The name of the Global datastore

applyImmediately :: Bool

This parameter causes the modifications in this request and any pending modifications to be applied, asynchronously and as soon as possible. Modifications to Global Replication Groups cannot be requested to be applied in PreferredMaintenceWindow.

newModifyGlobalReplicationGroup

Arguments

:: Text
$sel:globalReplicationGroupId:ModifyGlobalReplicationGroup'

-> Bool
$sel:applyImmediately:ModifyGlobalReplicationGroup'

-> ModifyGlobalReplicationGroup

Create a value of ModifyGlobalReplicationGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:automaticFailoverEnabled:ModifyGlobalReplicationGroup', modifyGlobalReplicationGroup_automaticFailoverEnabled - Determines whether a read replica is automatically promoted to read/write primary if the existing primary encounters a failure.

$sel:cacheParameterGroupName:ModifyGlobalReplicationGroup', modifyGlobalReplicationGroup_cacheParameterGroupName - The name of the cache parameter group to use with the Global datastore. It must be compatible with the major engine version used by the Global datastore.

$sel:engineVersion:ModifyGlobalReplicationGroup', modifyGlobalReplicationGroup_engineVersion - The upgraded version of the cache engine to be run on the clusters in the Global datastore.

$sel:cacheNodeType:ModifyGlobalReplicationGroup', modifyGlobalReplicationGroup_cacheNodeType - A valid cache node type that you want to scale this Global datastore to.

$sel:globalReplicationGroupDescription:ModifyGlobalReplicationGroup', modifyGlobalReplicationGroup_globalReplicationGroupDescription - A description of the Global datastore

$sel:globalReplicationGroupId:ModifyGlobalReplicationGroup', modifyGlobalReplicationGroup_globalReplicationGroupId - The name of the Global datastore

$sel:applyImmediately:ModifyGlobalReplicationGroup', modifyGlobalReplicationGroup_applyImmediately - This parameter causes the modifications in this request and any pending modifications to be applied, asynchronously and as soon as possible. Modifications to Global Replication Groups cannot be requested to be applied in PreferredMaintenceWindow.

Request Lenses

modifyGlobalReplicationGroup_automaticFailoverEnabled :: Lens' ModifyGlobalReplicationGroup (Maybe Bool)

Determines whether a read replica is automatically promoted to read/write primary if the existing primary encounters a failure.

modifyGlobalReplicationGroup_cacheParameterGroupName :: Lens' ModifyGlobalReplicationGroup (Maybe Text)

The name of the cache parameter group to use with the Global datastore. It must be compatible with the major engine version used by the Global datastore.

modifyGlobalReplicationGroup_engineVersion :: Lens' ModifyGlobalReplicationGroup (Maybe Text)

The upgraded version of the cache engine to be run on the clusters in the Global datastore.

modifyGlobalReplicationGroup_cacheNodeType :: Lens' ModifyGlobalReplicationGroup (Maybe Text)

A valid cache node type that you want to scale this Global datastore to.

modifyGlobalReplicationGroup_globalReplicationGroupDescription :: Lens' ModifyGlobalReplicationGroup (Maybe Text)

A description of the Global datastore

modifyGlobalReplicationGroup_globalReplicationGroupId :: Lens' ModifyGlobalReplicationGroup Text

The name of the Global datastore

modifyGlobalReplicationGroup_applyImmediately :: Lens' ModifyGlobalReplicationGroup Bool

This parameter causes the modifications in this request and any pending modifications to be applied, asynchronously and as soon as possible. Modifications to Global Replication Groups cannot be requested to be applied in PreferredMaintenceWindow.

Destructuring the Response

data ModifyGlobalReplicationGroupResponse

See: newModifyGlobalReplicationGroupResponse smart constructor.

Constructors

ModifyGlobalReplicationGroupResponse'

Fields

globalReplicationGroup :: Maybe GlobalReplicationGroup

httpStatus :: Int

The response's http status code.

newModifyGlobalReplicationGroupResponse

Arguments

:: Int
$sel:httpStatus:ModifyGlobalReplicationGroupResponse'

-> ModifyGlobalReplicationGroupResponse

Create a value of ModifyGlobalReplicationGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:globalReplicationGroup:ModifyGlobalReplicationGroupResponse', modifyGlobalReplicationGroupResponse_globalReplicationGroup - Undocumented member.

$sel:httpStatus:ModifyGlobalReplicationGroupResponse', modifyGlobalReplicationGroupResponse_httpStatus - The response's http status code.

Response Lenses

modifyGlobalReplicationGroupResponse_globalReplicationGroup :: Lens' ModifyGlobalReplicationGroupResponse (Maybe GlobalReplicationGroup)

Undocumented member.

modifyGlobalReplicationGroupResponse_httpStatus :: Lens' ModifyGlobalReplicationGroupResponse Int

The response's http status code.