Network.AWS.ElastiCache.ModifyReplicationGroup

Description

Modifies the settings for a replication group.

Scaling for Amazon ElastiCache for Redis (cluster mode enabled) in the ElastiCache User Guide

ModifyReplicationGroupShardConfiguration in the ElastiCache API Reference

This operation is valid for Redis only.

Creating a Request

data ModifyReplicationGroup

Represents the input of a ModifyReplicationGroups operation.

See: newModifyReplicationGroup smart constructor.

Constructors

ModifyReplicationGroup'

Fields

securityGroupIds :: Maybe [Text]

Specifies the VPC Security Groups associated with the clusters in the replication group.

This parameter can be used only with replication group containing clusters running in an Amazon Virtual Private Cloud (Amazon VPC).

automaticFailoverEnabled :: Maybe Bool

Determines whether a read replica is automatically promoted to read/write primary if the existing primary encounters a failure.

Valid values: true | false

cacheSecurityGroupNames :: Maybe [Text]

A list of cache security group names to authorize for the clusters in this replication group. This change is asynchronously applied as soon as possible.

This parameter can be used only with replication group containing clusters running outside of an Amazon Virtual Private Cloud (Amazon VPC).

Constraints: Must contain no more than 255 alphanumeric characters. Must not be Default.

authTokenUpdateStrategy :: Maybe AuthTokenUpdateStrategyType

Specifies the strategy to use to update the AUTH token. This parameter must be specified with the auth-token parameter. Possible values:

Rotate

Set

For more information, see Authenticating Users with Redis AUTH

primaryClusterId :: Maybe Text

For replication groups with a single primary, if this parameter is specified, ElastiCache promotes the specified cluster in the specified replication group to the primary role. The nodes of all other clusters in the replication group are read replicas.

snapshotWindow :: Maybe Text

The daily time range (in UTC) during which ElastiCache begins taking a daily snapshot of the node group (shard) specified by SnapshottingClusterId.

Example: 05:00-09:00

If you do not specify this parameter, ElastiCache automatically chooses an appropriate time range.

notificationTopicStatus :: Maybe Text

The status of the Amazon SNS notification topic for the replication group. Notifications are sent only if the status is active.

Valid values: active | inactive

cacheParameterGroupName :: Maybe Text

The name of the cache parameter group to apply to all of the clusters in this replication group. This change is asynchronously applied as soon as possible for parameters when the ApplyImmediately parameter is specified as true for this request.

userGroupIdsToRemove :: Maybe [Text]

The ID of the user group to disassociate from the replication group, meaning the users in the group no longer can access the replication group.

replicationGroupDescription :: Maybe Text

A description for the replication group. Maximum length is 255 characters.

snapshotRetentionLimit :: Maybe Int

The number of days for which ElastiCache retains automatic node group (shard) snapshots before deleting them. For example, if you set SnapshotRetentionLimit to 5, a snapshot that was taken today is retained for 5 days before being deleted.

Important If the value of SnapshotRetentionLimit is set to zero (0), backups are turned off.

userGroupIdsToAdd :: Maybe [Text]

The ID of the user group you are associating with the replication group.

nodeGroupId :: Maybe Text

Deprecated. This parameter is not used.

snapshottingClusterId :: Maybe Text

The cluster ID that is used as the daily snapshot source for the replication group. This parameter cannot be set for Redis (cluster mode enabled) replication groups.

multiAZEnabled :: Maybe Bool

A flag to indicate MultiAZ is enabled.

engineVersion :: Maybe Text

The upgraded version of the cache engine to be run on the clusters in the replication group.

Important: You can upgrade to a newer engine version (see Selecting a Cache Engine and Version), but you cannot downgrade to an earlier engine version. If you want to use an earlier engine version, you must delete the existing replication group and create it anew with the earlier engine version.

cacheNodeType :: Maybe Text

A valid cache node type that you want to scale this replication group to.

preferredMaintenanceWindow :: Maybe Text

Specifies the weekly time range during which maintenance on the cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period.

Valid values for ddd are:

sun

mon

tue

wed

thu

fri

sat

Example: sun:23:00-mon:01:30

notificationTopicArn :: Maybe Text

The Amazon Resource Name (ARN) of the Amazon SNS topic to which notifications are sent.

The Amazon SNS topic owner must be same as the replication group owner.

removeUserGroups :: Maybe Bool

Removes the user group associated with this replication group.

authToken :: Maybe Text

Reserved parameter. The password used to access a password protected server. This parameter must be specified with the auth-token-update-strategy parameter. Password constraints:

Must be only printable ASCII characters

Must be at least 16 characters and no more than 128 characters in length

Cannot contain any of the following characters: '/', '"', or '@', '%'

For more information, see AUTH password at AUTH.

logDeliveryConfigurations :: Maybe [LogDeliveryConfigurationRequest]

Specifies the destination, format and type of the logs.

applyImmediately :: Maybe Bool

If true, this parameter causes the modifications in this request and any pending modifications to be applied, asynchronously and as soon as possible, regardless of the PreferredMaintenanceWindow setting for the replication group.

If false, changes to the nodes in the replication group are applied on the next maintenance reboot, or the next failure reboot, whichever occurs first.

Valid values: true | false

Default: false

autoMinorVersionUpgrade :: Maybe Bool

This parameter is currently disabled.

replicationGroupId :: Text

The identifier of the replication group to modify.

newModifyReplicationGroup

Arguments

:: Text
$sel:replicationGroupId:ModifyReplicationGroup'

-> ModifyReplicationGroup

Create a value of ModifyReplicationGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:securityGroupIds:ModifyReplicationGroup', modifyReplicationGroup_securityGroupIds - Specifies the VPC Security Groups associated with the clusters in the replication group.

This parameter can be used only with replication group containing clusters running in an Amazon Virtual Private Cloud (Amazon VPC).

$sel:automaticFailoverEnabled:ModifyReplicationGroup', modifyReplicationGroup_automaticFailoverEnabled - Determines whether a read replica is automatically promoted to read/write primary if the existing primary encounters a failure.

Valid values: true | false

$sel:cacheSecurityGroupNames:ModifyReplicationGroup', modifyReplicationGroup_cacheSecurityGroupNames - A list of cache security group names to authorize for the clusters in this replication group. This change is asynchronously applied as soon as possible.

This parameter can be used only with replication group containing clusters running outside of an Amazon Virtual Private Cloud (Amazon VPC).

Constraints: Must contain no more than 255 alphanumeric characters. Must not be Default.

$sel:authTokenUpdateStrategy:ModifyReplicationGroup', modifyReplicationGroup_authTokenUpdateStrategy - Specifies the strategy to use to update the AUTH token. This parameter must be specified with the auth-token parameter. Possible values:

Rotate

Set

For more information, see Authenticating Users with Redis AUTH

$sel:primaryClusterId:ModifyReplicationGroup', modifyReplicationGroup_primaryClusterId - For replication groups with a single primary, if this parameter is specified, ElastiCache promotes the specified cluster in the specified replication group to the primary role. The nodes of all other clusters in the replication group are read replicas.

$sel:snapshotWindow:ModifyReplicationGroup', modifyReplicationGroup_snapshotWindow - The daily time range (in UTC) during which ElastiCache begins taking a daily snapshot of the node group (shard) specified by SnapshottingClusterId.

Example: 05:00-09:00

If you do not specify this parameter, ElastiCache automatically chooses an appropriate time range.

$sel:notificationTopicStatus:ModifyReplicationGroup', modifyReplicationGroup_notificationTopicStatus - The status of the Amazon SNS notification topic for the replication group. Notifications are sent only if the status is active.

Valid values: active | inactive

$sel:cacheParameterGroupName:ModifyReplicationGroup', modifyReplicationGroup_cacheParameterGroupName - The name of the cache parameter group to apply to all of the clusters in this replication group. This change is asynchronously applied as soon as possible for parameters when the ApplyImmediately parameter is specified as true for this request.

$sel:userGroupIdsToRemove:ModifyReplicationGroup', modifyReplicationGroup_userGroupIdsToRemove - The ID of the user group to disassociate from the replication group, meaning the users in the group no longer can access the replication group.

$sel:replicationGroupDescription:ModifyReplicationGroup', modifyReplicationGroup_replicationGroupDescription - A description for the replication group. Maximum length is 255 characters.

$sel:snapshotRetentionLimit:ModifyReplicationGroup', modifyReplicationGroup_snapshotRetentionLimit - The number of days for which ElastiCache retains automatic node group (shard) snapshots before deleting them. For example, if you set SnapshotRetentionLimit to 5, a snapshot that was taken today is retained for 5 days before being deleted.

Important If the value of SnapshotRetentionLimit is set to zero (0), backups are turned off.

$sel:userGroupIdsToAdd:ModifyReplicationGroup', modifyReplicationGroup_userGroupIdsToAdd - The ID of the user group you are associating with the replication group.

$sel:nodeGroupId:ModifyReplicationGroup', modifyReplicationGroup_nodeGroupId - Deprecated. This parameter is not used.

$sel:snapshottingClusterId:ModifyReplicationGroup', modifyReplicationGroup_snapshottingClusterId - The cluster ID that is used as the daily snapshot source for the replication group. This parameter cannot be set for Redis (cluster mode enabled) replication groups.

$sel:multiAZEnabled:ModifyReplicationGroup', modifyReplicationGroup_multiAZEnabled - A flag to indicate MultiAZ is enabled.

$sel:engineVersion:ModifyReplicationGroup', modifyReplicationGroup_engineVersion - The upgraded version of the cache engine to be run on the clusters in the replication group.

Important: You can upgrade to a newer engine version (see Selecting a Cache Engine and Version), but you cannot downgrade to an earlier engine version. If you want to use an earlier engine version, you must delete the existing replication group and create it anew with the earlier engine version.

$sel:cacheNodeType:ModifyReplicationGroup', modifyReplicationGroup_cacheNodeType - A valid cache node type that you want to scale this replication group to.

$sel:preferredMaintenanceWindow:ModifyReplicationGroup', modifyReplicationGroup_preferredMaintenanceWindow - Specifies the weekly time range during which maintenance on the cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period.

Valid values for ddd are:

sun

mon

tue

wed

thu

fri

sat

Example: sun:23:00-mon:01:30

$sel:notificationTopicArn:ModifyReplicationGroup', modifyReplicationGroup_notificationTopicArn - The Amazon Resource Name (ARN) of the Amazon SNS topic to which notifications are sent.

The Amazon SNS topic owner must be same as the replication group owner.

$sel:removeUserGroups:ModifyReplicationGroup', modifyReplicationGroup_removeUserGroups - Removes the user group associated with this replication group.

$sel:authToken:ModifyReplicationGroup', modifyReplicationGroup_authToken - Reserved parameter. The password used to access a password protected server. This parameter must be specified with the auth-token-update-strategy parameter. Password constraints:

Must be only printable ASCII characters

Must be at least 16 characters and no more than 128 characters in length

Cannot contain any of the following characters: '/', '"', or '@', '%'

For more information, see AUTH password at AUTH.

$sel:logDeliveryConfigurations:ModifyReplicationGroup', modifyReplicationGroup_logDeliveryConfigurations - Specifies the destination, format and type of the logs.

$sel:applyImmediately:ModifyReplicationGroup', modifyReplicationGroup_applyImmediately - If true, this parameter causes the modifications in this request and any pending modifications to be applied, asynchronously and as soon as possible, regardless of the PreferredMaintenanceWindow setting for the replication group.

If false, changes to the nodes in the replication group are applied on the next maintenance reboot, or the next failure reboot, whichever occurs first.

Valid values: true | false

Default: false

$sel:autoMinorVersionUpgrade:ModifyReplicationGroup', modifyReplicationGroup_autoMinorVersionUpgrade - This parameter is currently disabled.

$sel:replicationGroupId:ModifyReplicationGroup', modifyReplicationGroup_replicationGroupId - The identifier of the replication group to modify.

Request Lenses

modifyReplicationGroup_securityGroupIds :: Lens' ModifyReplicationGroup (Maybe [Text])

Specifies the VPC Security Groups associated with the clusters in the replication group.

This parameter can be used only with replication group containing clusters running in an Amazon Virtual Private Cloud (Amazon VPC).

modifyReplicationGroup_automaticFailoverEnabled :: Lens' ModifyReplicationGroup (Maybe Bool)

Determines whether a read replica is automatically promoted to read/write primary if the existing primary encounters a failure.

Valid values: true | false

modifyReplicationGroup_cacheSecurityGroupNames :: Lens' ModifyReplicationGroup (Maybe [Text])

A list of cache security group names to authorize for the clusters in this replication group. This change is asynchronously applied as soon as possible.

This parameter can be used only with replication group containing clusters running outside of an Amazon Virtual Private Cloud (Amazon VPC).

Constraints: Must contain no more than 255 alphanumeric characters. Must not be Default.

modifyReplicationGroup_authTokenUpdateStrategy :: Lens' ModifyReplicationGroup (Maybe AuthTokenUpdateStrategyType)

Specifies the strategy to use to update the AUTH token. This parameter must be specified with the auth-token parameter. Possible values:

Rotate

Set

For more information, see Authenticating Users with Redis AUTH

modifyReplicationGroup_primaryClusterId :: Lens' ModifyReplicationGroup (Maybe Text)

For replication groups with a single primary, if this parameter is specified, ElastiCache promotes the specified cluster in the specified replication group to the primary role. The nodes of all other clusters in the replication group are read replicas.

modifyReplicationGroup_snapshotWindow :: Lens' ModifyReplicationGroup (Maybe Text)

The daily time range (in UTC) during which ElastiCache begins taking a daily snapshot of the node group (shard) specified by SnapshottingClusterId.

Example: 05:00-09:00

If you do not specify this parameter, ElastiCache automatically chooses an appropriate time range.

modifyReplicationGroup_notificationTopicStatus :: Lens' ModifyReplicationGroup (Maybe Text)

The status of the Amazon SNS notification topic for the replication group. Notifications are sent only if the status is active.

Valid values: active | inactive

modifyReplicationGroup_cacheParameterGroupName :: Lens' ModifyReplicationGroup (Maybe Text)

The name of the cache parameter group to apply to all of the clusters in this replication group. This change is asynchronously applied as soon as possible for parameters when the ApplyImmediately parameter is specified as true for this request.

modifyReplicationGroup_userGroupIdsToRemove :: Lens' ModifyReplicationGroup (Maybe [Text])

The ID of the user group to disassociate from the replication group, meaning the users in the group no longer can access the replication group.

modifyReplicationGroup_replicationGroupDescription :: Lens' ModifyReplicationGroup (Maybe Text)

A description for the replication group. Maximum length is 255 characters.

modifyReplicationGroup_snapshotRetentionLimit :: Lens' ModifyReplicationGroup (Maybe Int)

The number of days for which ElastiCache retains automatic node group (shard) snapshots before deleting them. For example, if you set SnapshotRetentionLimit to 5, a snapshot that was taken today is retained for 5 days before being deleted.

Important If the value of SnapshotRetentionLimit is set to zero (0), backups are turned off.

modifyReplicationGroup_userGroupIdsToAdd :: Lens' ModifyReplicationGroup (Maybe [Text])

The ID of the user group you are associating with the replication group.

modifyReplicationGroup_nodeGroupId :: Lens' ModifyReplicationGroup (Maybe Text)

Deprecated. This parameter is not used.

modifyReplicationGroup_snapshottingClusterId :: Lens' ModifyReplicationGroup (Maybe Text)

The cluster ID that is used as the daily snapshot source for the replication group. This parameter cannot be set for Redis (cluster mode enabled) replication groups.

modifyReplicationGroup_multiAZEnabled :: Lens' ModifyReplicationGroup (Maybe Bool)

A flag to indicate MultiAZ is enabled.

modifyReplicationGroup_engineVersion :: Lens' ModifyReplicationGroup (Maybe Text)

The upgraded version of the cache engine to be run on the clusters in the replication group.

Important: You can upgrade to a newer engine version (see Selecting a Cache Engine and Version), but you cannot downgrade to an earlier engine version. If you want to use an earlier engine version, you must delete the existing replication group and create it anew with the earlier engine version.

modifyReplicationGroup_cacheNodeType :: Lens' ModifyReplicationGroup (Maybe Text)

A valid cache node type that you want to scale this replication group to.

modifyReplicationGroup_preferredMaintenanceWindow :: Lens' ModifyReplicationGroup (Maybe Text)

Specifies the weekly time range during which maintenance on the cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period.

Valid values for ddd are:

sun

mon

tue

wed

thu

fri

sat

Example: sun:23:00-mon:01:30

modifyReplicationGroup_notificationTopicArn :: Lens' ModifyReplicationGroup (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon SNS topic to which notifications are sent.

The Amazon SNS topic owner must be same as the replication group owner.

modifyReplicationGroup_removeUserGroups :: Lens' ModifyReplicationGroup (Maybe Bool)

Removes the user group associated with this replication group.

modifyReplicationGroup_authToken :: Lens' ModifyReplicationGroup (Maybe Text)

Reserved parameter. The password used to access a password protected server. This parameter must be specified with the auth-token-update-strategy parameter. Password constraints:

Must be only printable ASCII characters

Must be at least 16 characters and no more than 128 characters in length

Cannot contain any of the following characters: '/', '"', or '@', '%'

For more information, see AUTH password at AUTH.

modifyReplicationGroup_logDeliveryConfigurations :: Lens' ModifyReplicationGroup (Maybe [LogDeliveryConfigurationRequest])

Specifies the destination, format and type of the logs.

modifyReplicationGroup_applyImmediately :: Lens' ModifyReplicationGroup (Maybe Bool)

If true, this parameter causes the modifications in this request and any pending modifications to be applied, asynchronously and as soon as possible, regardless of the PreferredMaintenanceWindow setting for the replication group.

If false, changes to the nodes in the replication group are applied on the next maintenance reboot, or the next failure reboot, whichever occurs first.

Valid values: true | false

Default: false

modifyReplicationGroup_autoMinorVersionUpgrade :: Lens' ModifyReplicationGroup (Maybe Bool)

This parameter is currently disabled.

modifyReplicationGroup_replicationGroupId :: Lens' ModifyReplicationGroup Text

The identifier of the replication group to modify.

Destructuring the Response

data ModifyReplicationGroupResponse

See: newModifyReplicationGroupResponse smart constructor.

Constructors

ModifyReplicationGroupResponse'

Fields

replicationGroup :: Maybe ReplicationGroup

httpStatus :: Int

The response's http status code.

newModifyReplicationGroupResponse

Arguments

:: Int
$sel:httpStatus:ModifyReplicationGroupResponse'

-> ModifyReplicationGroupResponse

Create a value of ModifyReplicationGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:replicationGroup:ModifyReplicationGroupResponse', modifyReplicationGroupResponse_replicationGroup - Undocumented member.

$sel:httpStatus:ModifyReplicationGroupResponse', modifyReplicationGroupResponse_httpStatus - The response's http status code.

Response Lenses

modifyReplicationGroupResponse_replicationGroup :: Lens' ModifyReplicationGroupResponse (Maybe ReplicationGroup)

Undocumented member.

modifyReplicationGroupResponse_httpStatus :: Lens' ModifyReplicationGroupResponse Int

The response's http status code.