Network.AWS.ElastiCache.ModifyUserGroup

Description

Changes the list of users that belong to the user group.

Creating a Request

data ModifyUserGroup

See: newModifyUserGroup smart constructor.

Constructors

ModifyUserGroup'

Fields

userIdsToRemove :: Maybe (NonEmpty Text)

The list of user IDs to remove from the user group.

userIdsToAdd :: Maybe (NonEmpty Text)

The list of user IDs to add to the user group.

userGroupId :: Text

The ID of the user group.

newModifyUserGroup

Arguments

:: Text
$sel:userGroupId:ModifyUserGroup'

-> ModifyUserGroup

Create a value of ModifyUserGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userIdsToRemove:ModifyUserGroup', modifyUserGroup_userIdsToRemove - The list of user IDs to remove from the user group.

$sel:userIdsToAdd:ModifyUserGroup', modifyUserGroup_userIdsToAdd - The list of user IDs to add to the user group.

$sel:userGroupId:ModifyUserGroup', modifyUserGroup_userGroupId - The ID of the user group.

Request Lenses

modifyUserGroup_userIdsToRemove :: Lens' ModifyUserGroup (Maybe (NonEmpty Text))

The list of user IDs to remove from the user group.

modifyUserGroup_userIdsToAdd :: Lens' ModifyUserGroup (Maybe (NonEmpty Text))

The list of user IDs to add to the user group.

modifyUserGroup_userGroupId :: Lens' ModifyUserGroup Text

The ID of the user group.

Destructuring the Response

data UserGroup

See: newUserGroup smart constructor.

Constructors

UserGroup'

Fields

status :: Maybe Text

Indicates user group status. Can be "creating", "active", "modifying", "deleting".

replicationGroups :: Maybe [Text]

A list of replication groups that the user group can access.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the user group.

userIds :: Maybe [Text]

The list of user IDs that belong to the user group.

engine :: Maybe Text

The current supported value is Redis.

userGroupId :: Maybe Text

The ID of the user group.

pendingChanges :: Maybe UserGroupPendingChanges

A list of updates being applied to the user group.

newUserGroup :: UserGroup

Create a value of UserGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:UserGroup', userGroup_status - Indicates user group status. Can be "creating", "active", "modifying", "deleting".

$sel:replicationGroups:UserGroup', userGroup_replicationGroups - A list of replication groups that the user group can access.

$sel:arn:UserGroup', userGroup_arn - The Amazon Resource Name (ARN) of the user group.

$sel:userIds:UserGroup', userGroup_userIds - The list of user IDs that belong to the user group.

$sel:engine:UserGroup', userGroup_engine - The current supported value is Redis.

$sel:userGroupId:UserGroup', userGroup_userGroupId - The ID of the user group.

$sel:pendingChanges:UserGroup', userGroup_pendingChanges - A list of updates being applied to the user group.

Response Lenses

userGroup_status :: Lens' UserGroup (Maybe Text)

Indicates user group status. Can be "creating", "active", "modifying", "deleting".

userGroup_replicationGroups :: Lens' UserGroup (Maybe [Text])

A list of replication groups that the user group can access.

userGroup_arn :: Lens' UserGroup (Maybe Text)

The Amazon Resource Name (ARN) of the user group.

userGroup_userIds :: Lens' UserGroup (Maybe [Text])

The list of user IDs that belong to the user group.

userGroup_engine :: Lens' UserGroup (Maybe Text)

The current supported value is Redis.

userGroup_userGroupId :: Lens' UserGroup (Maybe Text)

The ID of the user group.

userGroup_pendingChanges :: Lens' UserGroup (Maybe UserGroupPendingChanges)

A list of updates being applied to the user group.