Network.AWS.ElastiCache.RebootCacheCluster

Description

Reboots some, or all, of the cache nodes within a provisioned cluster. This operation applies any modified cache parameter groups to the cluster. The reboot operation takes place as soon as possible, and results in a momentary outage to the cluster. During the reboot, the cluster status is set to REBOOTING.

The reboot causes the contents of the cache (for each cache node being rebooted) to be lost.

When the reboot is complete, a cluster event is created.

Rebooting a cluster is currently supported on Memcached and Redis (cluster mode disabled) clusters. Rebooting is not supported on Redis (cluster mode enabled) clusters.

If you make changes to parameters that require a Redis (cluster mode enabled) cluster reboot for the changes to be applied, see Rebooting a Cluster for an alternate process.

Creating a Request

data RebootCacheCluster

Represents the input of a RebootCacheCluster operation.

See: newRebootCacheCluster smart constructor.

Constructors

RebootCacheCluster'

Fields

cacheClusterId :: Text

The cluster identifier. This parameter is stored as a lowercase string.

cacheNodeIdsToReboot :: [Text]

A list of cache node IDs to reboot. A node ID is a numeric identifier (0001, 0002, etc.). To reboot an entire cluster, specify all of the cache node IDs.

newRebootCacheCluster

Arguments

:: Text
$sel:cacheClusterId:RebootCacheCluster'

-> RebootCacheCluster

Create a value of RebootCacheCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cacheClusterId:RebootCacheCluster', rebootCacheCluster_cacheClusterId - The cluster identifier. This parameter is stored as a lowercase string.

$sel:cacheNodeIdsToReboot:RebootCacheCluster', rebootCacheCluster_cacheNodeIdsToReboot - A list of cache node IDs to reboot. A node ID is a numeric identifier (0001, 0002, etc.). To reboot an entire cluster, specify all of the cache node IDs.

Request Lenses

rebootCacheCluster_cacheClusterId :: Lens' RebootCacheCluster Text

The cluster identifier. This parameter is stored as a lowercase string.

rebootCacheCluster_cacheNodeIdsToReboot :: Lens' RebootCacheCluster [Text]

A list of cache node IDs to reboot. A node ID is a numeric identifier (0001, 0002, etc.). To reboot an entire cluster, specify all of the cache node IDs.

Destructuring the Response

data RebootCacheClusterResponse

See: newRebootCacheClusterResponse smart constructor.

Constructors

RebootCacheClusterResponse'

Fields

cacheCluster :: Maybe CacheCluster

httpStatus :: Int

The response's http status code.

newRebootCacheClusterResponse

Arguments

:: Int
$sel:httpStatus:RebootCacheClusterResponse'

-> RebootCacheClusterResponse

Create a value of RebootCacheClusterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cacheCluster:RebootCacheClusterResponse', rebootCacheClusterResponse_cacheCluster - Undocumented member.

$sel:httpStatus:RebootCacheClusterResponse', rebootCacheClusterResponse_httpStatus - The response's http status code.

Response Lenses

rebootCacheClusterResponse_cacheCluster :: Lens' RebootCacheClusterResponse (Maybe CacheCluster)

Undocumented member.

rebootCacheClusterResponse_httpStatus :: Lens' RebootCacheClusterResponse Int

The response's http status code.