Network.AWS.ElastiCache.RevokeCacheSecurityGroupIngress

Description

Revokes ingress from a cache security group. Use this operation to disallow access from an Amazon EC2 security group that had been previously authorized.

Creating a Request

data RevokeCacheSecurityGroupIngress

Represents the input of a RevokeCacheSecurityGroupIngress operation.

See: newRevokeCacheSecurityGroupIngress smart constructor.

Constructors

RevokeCacheSecurityGroupIngress'

Fields

cacheSecurityGroupName :: Text

The name of the cache security group to revoke ingress from.

eC2SecurityGroupName :: Text

The name of the Amazon EC2 security group to revoke access from.

eC2SecurityGroupOwnerId :: Text

The Amazon account number of the Amazon EC2 security group owner. Note that this is not the same thing as an Amazon access key ID - you must provide a valid Amazon account number for this parameter.

newRevokeCacheSecurityGroupIngress

Arguments

:: Text
$sel:cacheSecurityGroupName:RevokeCacheSecurityGroupIngress'

-> Text
$sel:eC2SecurityGroupName:RevokeCacheSecurityGroupIngress'

-> Text
$sel:eC2SecurityGroupOwnerId:RevokeCacheSecurityGroupIngress'

-> RevokeCacheSecurityGroupIngress

Create a value of RevokeCacheSecurityGroupIngress with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cacheSecurityGroupName:RevokeCacheSecurityGroupIngress', revokeCacheSecurityGroupIngress_cacheSecurityGroupName - The name of the cache security group to revoke ingress from.

$sel:eC2SecurityGroupName:RevokeCacheSecurityGroupIngress', revokeCacheSecurityGroupIngress_eC2SecurityGroupName - The name of the Amazon EC2 security group to revoke access from.

$sel:eC2SecurityGroupOwnerId:RevokeCacheSecurityGroupIngress', revokeCacheSecurityGroupIngress_eC2SecurityGroupOwnerId - The Amazon account number of the Amazon EC2 security group owner. Note that this is not the same thing as an Amazon access key ID - you must provide a valid Amazon account number for this parameter.

Request Lenses

revokeCacheSecurityGroupIngress_cacheSecurityGroupName :: Lens' RevokeCacheSecurityGroupIngress Text

The name of the cache security group to revoke ingress from.

revokeCacheSecurityGroupIngress_eC2SecurityGroupName :: Lens' RevokeCacheSecurityGroupIngress Text

The name of the Amazon EC2 security group to revoke access from.

revokeCacheSecurityGroupIngress_eC2SecurityGroupOwnerId :: Lens' RevokeCacheSecurityGroupIngress Text

The Amazon account number of the Amazon EC2 security group owner. Note that this is not the same thing as an Amazon access key ID - you must provide a valid Amazon account number for this parameter.

Destructuring the Response

data RevokeCacheSecurityGroupIngressResponse

See: newRevokeCacheSecurityGroupIngressResponse smart constructor.

Constructors

RevokeCacheSecurityGroupIngressResponse'

Fields

cacheSecurityGroup :: Maybe CacheSecurityGroup

httpStatus :: Int

The response's http status code.

newRevokeCacheSecurityGroupIngressResponse

Arguments

:: Int
$sel:httpStatus:RevokeCacheSecurityGroupIngressResponse'

-> RevokeCacheSecurityGroupIngressResponse

Create a value of RevokeCacheSecurityGroupIngressResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cacheSecurityGroup:RevokeCacheSecurityGroupIngressResponse', revokeCacheSecurityGroupIngressResponse_cacheSecurityGroup - Undocumented member.

$sel:httpStatus:RevokeCacheSecurityGroupIngressResponse', revokeCacheSecurityGroupIngressResponse_httpStatus - The response's http status code.

Response Lenses

revokeCacheSecurityGroupIngressResponse_cacheSecurityGroup :: Lens' RevokeCacheSecurityGroupIngressResponse (Maybe CacheSecurityGroup)

Undocumented member.

revokeCacheSecurityGroupIngressResponse_httpStatus :: Lens' RevokeCacheSecurityGroupIngressResponse Int

The response's http status code.