Network.AWS.ElastiCache.TestFailover

Description

Represents the input of a TestFailover operation which test automatic failover on a specified node group (called shard in the console) in a replication group (called cluster in the console).

Note the following

A customer can use this operation to test automatic failover on up to 5 shards (called node groups in the ElastiCache API and Amazon CLI) in any rolling 24-hour period.

If calling this operation on shards in different clusters (called replication groups in the API and CLI), the calls can be made concurrently.

If calling this operation multiple times on different shards in the same Redis (cluster mode enabled) replication group, the first node replacement must complete before a subsequent call can be made.

To determine whether the node replacement is complete you can check Events using the Amazon ElastiCache console, the Amazon CLI, or the ElastiCache API. Look for the following automatic failover related events, listed here in order of occurrance:

Replication group message: Test Failover API called for node group <node-group-id>

Cache cluster message: Failover from primary node <primary-node-id> to replica node <node-id> completed

Replication group message: Failover from primary node <primary-node-id> to replica node <node-id> completed

Cache cluster message: Recovering cache nodes <node-id>

Cache cluster message: Finished recovery for cache nodes <node-id>

For more information see:

Viewing ElastiCache Events in the ElastiCache User Guide

DescribeEvents in the ElastiCache API Reference

Also see, Testing Multi-AZ in the ElastiCache User Guide.

Creating a Request

data TestFailover

See: newTestFailover smart constructor.

Constructors

TestFailover'

Fields

replicationGroupId :: Text

The name of the replication group (console: cluster) whose automatic failover is being tested by this operation.

nodeGroupId :: Text

The name of the node group (called shard in the console) in this replication group on which automatic failover is to be tested. You may test automatic failover on up to 5 node groups in any rolling 24-hour period.

newTestFailover

Arguments

:: Text
$sel:replicationGroupId:TestFailover'

-> Text
$sel:nodeGroupId:TestFailover'

-> TestFailover

Create a value of TestFailover with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:replicationGroupId:TestFailover', testFailover_replicationGroupId - The name of the replication group (console: cluster) whose automatic failover is being tested by this operation.

$sel:nodeGroupId:TestFailover', testFailover_nodeGroupId - The name of the node group (called shard in the console) in this replication group on which automatic failover is to be tested. You may test automatic failover on up to 5 node groups in any rolling 24-hour period.

Request Lenses

testFailover_replicationGroupId :: Lens' TestFailover Text

The name of the replication group (console: cluster) whose automatic failover is being tested by this operation.

testFailover_nodeGroupId :: Lens' TestFailover Text

The name of the node group (called shard in the console) in this replication group on which automatic failover is to be tested. You may test automatic failover on up to 5 node groups in any rolling 24-hour period.

Destructuring the Response

data TestFailoverResponse

See: newTestFailoverResponse smart constructor.

Constructors

TestFailoverResponse'

Fields

replicationGroup :: Maybe ReplicationGroup

httpStatus :: Int

The response's http status code.

newTestFailoverResponse

Arguments

:: Int
$sel:httpStatus:TestFailoverResponse'

-> TestFailoverResponse

Create a value of TestFailoverResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:replicationGroup:TestFailoverResponse', testFailoverResponse_replicationGroup - Undocumented member.

$sel:httpStatus:TestFailoverResponse', testFailoverResponse_httpStatus - The response's http status code.

Response Lenses

testFailoverResponse_replicationGroup :: Lens' TestFailoverResponse (Maybe ReplicationGroup)

Undocumented member.

testFailoverResponse_httpStatus :: Lens' TestFailoverResponse Int

The response's http status code.