Network.AWS.ElastiCache.Types.CacheEngineVersion

Description

Documentation

data CacheEngineVersion

Provides all of the details about a particular cache engine version.

See: newCacheEngineVersion smart constructor.

Constructors

CacheEngineVersion'

Fields

cacheEngineDescription :: Maybe Text

The description of the cache engine.

cacheEngineVersionDescription :: Maybe Text

The description of the cache engine version.

engineVersion :: Maybe Text

The version number of the cache engine.

cacheParameterGroupFamily :: Maybe Text

The name of the cache parameter group family associated with this cache engine.

Valid values are: memcached1.4 | memcached1.5 | memcached1.6 | redis2.6 | redis2.8 | redis3.2 | redis4.0 | redis5.0 | redis6.x |

engine :: Maybe Text

The name of the cache engine.

newCacheEngineVersion :: CacheEngineVersion

Create a value of CacheEngineVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cacheEngineDescription:CacheEngineVersion', cacheEngineVersion_cacheEngineDescription - The description of the cache engine.

$sel:cacheEngineVersionDescription:CacheEngineVersion', cacheEngineVersion_cacheEngineVersionDescription - The description of the cache engine version.

$sel:engineVersion:CacheEngineVersion', cacheEngineVersion_engineVersion - The version number of the cache engine.

$sel:cacheParameterGroupFamily:CacheEngineVersion', cacheEngineVersion_cacheParameterGroupFamily - The name of the cache parameter group family associated with this cache engine.

Valid values are: memcached1.4 | memcached1.5 | memcached1.6 | redis2.6 | redis2.8 | redis3.2 | redis4.0 | redis5.0 | redis6.x |

$sel:engine:CacheEngineVersion', cacheEngineVersion_engine - The name of the cache engine.

cacheEngineVersion_cacheEngineDescription :: Lens' CacheEngineVersion (Maybe Text)

The description of the cache engine.

cacheEngineVersion_cacheEngineVersionDescription :: Lens' CacheEngineVersion (Maybe Text)

The description of the cache engine version.

cacheEngineVersion_engineVersion :: Lens' CacheEngineVersion (Maybe Text)

The version number of the cache engine.

cacheEngineVersion_cacheParameterGroupFamily :: Lens' CacheEngineVersion (Maybe Text)

The name of the cache parameter group family associated with this cache engine.

Valid values are: memcached1.4 | memcached1.5 | memcached1.6 | redis2.6 | redis2.8 | redis3.2 | redis4.0 | redis5.0 | redis6.x |

cacheEngineVersion_engine :: Lens' CacheEngineVersion (Maybe Text)

The name of the cache engine.