Network.AWS.ElastiCache.Types.CacheNodeTypeSpecificParameter

Description

Documentation

data CacheNodeTypeSpecificParameter

A parameter that has a different value for each cache node type it is applied to. For example, in a Redis cluster, a cache.m1.large cache node type would have a larger maxmemory value than a cache.m1.small type.

See: newCacheNodeTypeSpecificParameter smart constructor.

Constructors

CacheNodeTypeSpecificParameter'

Fields

cacheNodeTypeSpecificValues :: Maybe [CacheNodeTypeSpecificValue]

A list of cache node types and their corresponding values for this parameter.

minimumEngineVersion :: Maybe Text

The earliest cache engine version to which the parameter can apply.

source :: Maybe Text

The source of the parameter value.

isModifiable :: Maybe Bool

Indicates whether (true) or not (false) the parameter can be modified. Some parameters have security or operational implications that prevent them from being changed.

dataType :: Maybe Text

The valid data type for the parameter.

allowedValues :: Maybe Text

The valid range of values for the parameter.

parameterName :: Maybe Text

The name of the parameter.

description :: Maybe Text

A description of the parameter.

changeType :: Maybe ChangeType

Indicates whether a change to the parameter is applied immediately or requires a reboot for the change to be applied. You can force a reboot or wait until the next maintenance window's reboot. For more information, see Rebooting a Cluster.

newCacheNodeTypeSpecificParameter :: CacheNodeTypeSpecificParameter

Create a value of CacheNodeTypeSpecificParameter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cacheNodeTypeSpecificValues:CacheNodeTypeSpecificParameter', cacheNodeTypeSpecificParameter_cacheNodeTypeSpecificValues - A list of cache node types and their corresponding values for this parameter.

$sel:minimumEngineVersion:CacheNodeTypeSpecificParameter', cacheNodeTypeSpecificParameter_minimumEngineVersion - The earliest cache engine version to which the parameter can apply.

$sel:source:CacheNodeTypeSpecificParameter', cacheNodeTypeSpecificParameter_source - The source of the parameter value.

$sel:isModifiable:CacheNodeTypeSpecificParameter', cacheNodeTypeSpecificParameter_isModifiable - Indicates whether (true) or not (false) the parameter can be modified. Some parameters have security or operational implications that prevent them from being changed.

$sel:dataType:CacheNodeTypeSpecificParameter', cacheNodeTypeSpecificParameter_dataType - The valid data type for the parameter.

$sel:allowedValues:CacheNodeTypeSpecificParameter', cacheNodeTypeSpecificParameter_allowedValues - The valid range of values for the parameter.

$sel:parameterName:CacheNodeTypeSpecificParameter', cacheNodeTypeSpecificParameter_parameterName - The name of the parameter.

$sel:description:CacheNodeTypeSpecificParameter', cacheNodeTypeSpecificParameter_description - A description of the parameter.

$sel:changeType:CacheNodeTypeSpecificParameter', cacheNodeTypeSpecificParameter_changeType - Indicates whether a change to the parameter is applied immediately or requires a reboot for the change to be applied. You can force a reboot or wait until the next maintenance window's reboot. For more information, see Rebooting a Cluster.

cacheNodeTypeSpecificParameter_cacheNodeTypeSpecificValues :: Lens' CacheNodeTypeSpecificParameter (Maybe [CacheNodeTypeSpecificValue])

A list of cache node types and their corresponding values for this parameter.

cacheNodeTypeSpecificParameter_minimumEngineVersion :: Lens' CacheNodeTypeSpecificParameter (Maybe Text)

The earliest cache engine version to which the parameter can apply.

cacheNodeTypeSpecificParameter_source :: Lens' CacheNodeTypeSpecificParameter (Maybe Text)

The source of the parameter value.

cacheNodeTypeSpecificParameter_isModifiable :: Lens' CacheNodeTypeSpecificParameter (Maybe Bool)

Indicates whether (true) or not (false) the parameter can be modified. Some parameters have security or operational implications that prevent them from being changed.

cacheNodeTypeSpecificParameter_dataType :: Lens' CacheNodeTypeSpecificParameter (Maybe Text)

The valid data type for the parameter.

cacheNodeTypeSpecificParameter_allowedValues :: Lens' CacheNodeTypeSpecificParameter (Maybe Text)

The valid range of values for the parameter.

cacheNodeTypeSpecificParameter_parameterName :: Lens' CacheNodeTypeSpecificParameter (Maybe Text)

The name of the parameter.

cacheNodeTypeSpecificParameter_description :: Lens' CacheNodeTypeSpecificParameter (Maybe Text)

A description of the parameter.

cacheNodeTypeSpecificParameter_changeType :: Lens' CacheNodeTypeSpecificParameter (Maybe ChangeType)

Indicates whether a change to the parameter is applied immediately or requires a reboot for the change to be applied. You can force a reboot or wait until the next maintenance window's reboot. For more information, see Rebooting a Cluster.