Network.AWS.ElastiCache.Types.CacheNodeTypeSpecificValue

Description

Documentation

data CacheNodeTypeSpecificValue

A value that applies only to a certain cache node type.

See: newCacheNodeTypeSpecificValue smart constructor.

Constructors

CacheNodeTypeSpecificValue'

Fields

cacheNodeType :: Maybe Text

The cache node type for which this value applies.

value :: Maybe Text

The value for the cache node type.

newCacheNodeTypeSpecificValue :: CacheNodeTypeSpecificValue

Create a value of CacheNodeTypeSpecificValue with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cacheNodeType:CacheNodeTypeSpecificValue', cacheNodeTypeSpecificValue_cacheNodeType - The cache node type for which this value applies.

$sel:value:CacheNodeTypeSpecificValue', cacheNodeTypeSpecificValue_value - The value for the cache node type.

cacheNodeTypeSpecificValue_cacheNodeType :: Lens' CacheNodeTypeSpecificValue (Maybe Text)

The cache node type for which this value applies.

cacheNodeTypeSpecificValue_value :: Lens' CacheNodeTypeSpecificValue (Maybe Text)

The value for the cache node type.