Network.AWS.ElastiCache.Types.CacheNodeUpdateStatus

Description

Documentation

data CacheNodeUpdateStatus

The status of the service update on the cache node

See: newCacheNodeUpdateStatus smart constructor.

Constructors

CacheNodeUpdateStatus'

Fields

nodeUpdateStatusModifiedDate :: Maybe ISO8601

The date when the NodeUpdateStatus was last modified>

nodeUpdateInitiatedBy :: Maybe NodeUpdateInitiatedBy

Reflects whether the update was initiated by the customer or automatically applied

nodeUpdateStatus :: Maybe NodeUpdateStatus

The update status of the node

nodeUpdateInitiatedDate :: Maybe ISO8601

The date when the update is triggered

cacheNodeId :: Maybe Text

The node ID of the cache cluster

nodeDeletionDate :: Maybe ISO8601

The deletion date of the node

nodeUpdateStartDate :: Maybe ISO8601

The start date of the update for a node

nodeUpdateEndDate :: Maybe ISO8601

The end date of the update for a node

newCacheNodeUpdateStatus :: CacheNodeUpdateStatus

Create a value of CacheNodeUpdateStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nodeUpdateStatusModifiedDate:CacheNodeUpdateStatus', cacheNodeUpdateStatus_nodeUpdateStatusModifiedDate - The date when the NodeUpdateStatus was last modified>

$sel:nodeUpdateInitiatedBy:CacheNodeUpdateStatus', cacheNodeUpdateStatus_nodeUpdateInitiatedBy - Reflects whether the update was initiated by the customer or automatically applied

$sel:nodeUpdateStatus:CacheNodeUpdateStatus', cacheNodeUpdateStatus_nodeUpdateStatus - The update status of the node

$sel:nodeUpdateInitiatedDate:CacheNodeUpdateStatus', cacheNodeUpdateStatus_nodeUpdateInitiatedDate - The date when the update is triggered

$sel:cacheNodeId:CacheNodeUpdateStatus', cacheNodeUpdateStatus_cacheNodeId - The node ID of the cache cluster

$sel:nodeDeletionDate:CacheNodeUpdateStatus', cacheNodeUpdateStatus_nodeDeletionDate - The deletion date of the node

$sel:nodeUpdateStartDate:CacheNodeUpdateStatus', cacheNodeUpdateStatus_nodeUpdateStartDate - The start date of the update for a node

$sel:nodeUpdateEndDate:CacheNodeUpdateStatus', cacheNodeUpdateStatus_nodeUpdateEndDate - The end date of the update for a node

cacheNodeUpdateStatus_nodeUpdateStatusModifiedDate :: Lens' CacheNodeUpdateStatus (Maybe UTCTime)

The date when the NodeUpdateStatus was last modified>

cacheNodeUpdateStatus_nodeUpdateInitiatedBy :: Lens' CacheNodeUpdateStatus (Maybe NodeUpdateInitiatedBy)

Reflects whether the update was initiated by the customer or automatically applied

cacheNodeUpdateStatus_nodeUpdateStatus :: Lens' CacheNodeUpdateStatus (Maybe NodeUpdateStatus)

The update status of the node

cacheNodeUpdateStatus_nodeUpdateInitiatedDate :: Lens' CacheNodeUpdateStatus (Maybe UTCTime)

The date when the update is triggered

cacheNodeUpdateStatus_cacheNodeId :: Lens' CacheNodeUpdateStatus (Maybe Text)

The node ID of the cache cluster

cacheNodeUpdateStatus_nodeDeletionDate :: Lens' CacheNodeUpdateStatus (Maybe UTCTime)

The deletion date of the node

cacheNodeUpdateStatus_nodeUpdateStartDate :: Lens' CacheNodeUpdateStatus (Maybe UTCTime)

The start date of the update for a node

cacheNodeUpdateStatus_nodeUpdateEndDate :: Lens' CacheNodeUpdateStatus (Maybe UTCTime)

The end date of the update for a node