Network.AWS.ElastiCache.Types.CacheParameterGroup

Description

Documentation

data CacheParameterGroup

Represents the output of a CreateCacheParameterGroup operation.

See: newCacheParameterGroup smart constructor.

Constructors

CacheParameterGroup'

Fields

cacheParameterGroupFamily :: Maybe Text

The name of the cache parameter group family that this cache parameter group is compatible with.

Valid values are: memcached1.4 | memcached1.5 | memcached1.6 | redis2.6 | redis2.8 | redis3.2 | redis4.0 | redis5.0 | redis6.x |

arn :: Maybe Text

The ARN (Amazon Resource Name) of the cache parameter group.

cacheParameterGroupName :: Maybe Text

The name of the cache parameter group.

isGlobal :: Maybe Bool

Indicates whether the parameter group is associated with a Global datastore

description :: Maybe Text

The description for this cache parameter group.

newCacheParameterGroup :: CacheParameterGroup

Create a value of CacheParameterGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cacheParameterGroupFamily:CacheParameterGroup', cacheParameterGroup_cacheParameterGroupFamily - The name of the cache parameter group family that this cache parameter group is compatible with.

Valid values are: memcached1.4 | memcached1.5 | memcached1.6 | redis2.6 | redis2.8 | redis3.2 | redis4.0 | redis5.0 | redis6.x |

$sel:arn:CacheParameterGroup', cacheParameterGroup_arn - The ARN (Amazon Resource Name) of the cache parameter group.

$sel:cacheParameterGroupName:CacheParameterGroup', cacheParameterGroup_cacheParameterGroupName - The name of the cache parameter group.

$sel:isGlobal:CacheParameterGroup', cacheParameterGroup_isGlobal - Indicates whether the parameter group is associated with a Global datastore

$sel:description:CacheParameterGroup', cacheParameterGroup_description - The description for this cache parameter group.

cacheParameterGroup_cacheParameterGroupFamily :: Lens' CacheParameterGroup (Maybe Text)

The name of the cache parameter group family that this cache parameter group is compatible with.

Valid values are: memcached1.4 | memcached1.5 | memcached1.6 | redis2.6 | redis2.8 | redis3.2 | redis4.0 | redis5.0 | redis6.x |

cacheParameterGroup_arn :: Lens' CacheParameterGroup (Maybe Text)

The ARN (Amazon Resource Name) of the cache parameter group.

cacheParameterGroup_cacheParameterGroupName :: Lens' CacheParameterGroup (Maybe Text)

The name of the cache parameter group.

cacheParameterGroup_isGlobal :: Lens' CacheParameterGroup (Maybe Bool)

Indicates whether the parameter group is associated with a Global datastore

cacheParameterGroup_description :: Lens' CacheParameterGroup (Maybe Text)

The description for this cache parameter group.