Network.AWS.ElastiCache.Types.ConfigureShard

Description

Documentation

data ConfigureShard

Node group (shard) configuration options when adding or removing replicas. Each node group (shard) configuration has the following members: NodeGroupId, NewReplicaCount, and PreferredAvailabilityZones.

See: newConfigureShard smart constructor.

Constructors

ConfigureShard'

Fields

preferredAvailabilityZones :: Maybe [Text]

A list of PreferredAvailabilityZone strings that specify which availability zones the replication group's nodes are to be in. The nummber of PreferredAvailabilityZone values must equal the value of NewReplicaCount plus 1 to account for the primary node. If this member of ReplicaConfiguration is omitted, ElastiCache for Redis selects the availability zone for each of the replicas.

preferredOutpostArns :: Maybe [Text]

The outpost ARNs in which the cache cluster is created.

nodeGroupId :: Text

The 4-digit id for the node group you are configuring. For Redis (cluster mode disabled) replication groups, the node group id is always 0001. To find a Redis (cluster mode enabled)'s node group's (shard's) id, see Finding a Shard's Id.

newReplicaCount' :: Int

The number of replicas you want in this node group at the end of this operation. The maximum value for NewReplicaCount is 5. The minimum value depends upon the type of Redis replication group you are working with.

The minimum number of replicas in a shard or replication group is:

Redis (cluster mode disabled)

If Multi-AZ: 1

If Multi-AZ: 0

Redis (cluster mode enabled): 0 (though you will not be able to failover to a replica if your primary node fails)

newConfigureShard

Arguments

:: Text
$sel:nodeGroupId:ConfigureShard'

-> Int
$sel:newReplicaCount':ConfigureShard'

-> ConfigureShard

Create a value of ConfigureShard with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:preferredAvailabilityZones:ConfigureShard', configureShard_preferredAvailabilityZones - A list of PreferredAvailabilityZone strings that specify which availability zones the replication group's nodes are to be in. The nummber of PreferredAvailabilityZone values must equal the value of NewReplicaCount plus 1 to account for the primary node. If this member of ReplicaConfiguration is omitted, ElastiCache for Redis selects the availability zone for each of the replicas.

$sel:preferredOutpostArns:ConfigureShard', configureShard_preferredOutpostArns - The outpost ARNs in which the cache cluster is created.

$sel:nodeGroupId:ConfigureShard', configureShard_nodeGroupId - The 4-digit id for the node group you are configuring. For Redis (cluster mode disabled) replication groups, the node group id is always 0001. To find a Redis (cluster mode enabled)'s node group's (shard's) id, see Finding a Shard's Id.

$sel:newReplicaCount':ConfigureShard', configureShard_newReplicaCount - The number of replicas you want in this node group at the end of this operation. The maximum value for NewReplicaCount is 5. The minimum value depends upon the type of Redis replication group you are working with.

The minimum number of replicas in a shard or replication group is:

Redis (cluster mode disabled)

If Multi-AZ: 1

If Multi-AZ: 0

Redis (cluster mode enabled): 0 (though you will not be able to failover to a replica if your primary node fails)

configureShard_preferredAvailabilityZones :: Lens' ConfigureShard (Maybe [Text])

A list of PreferredAvailabilityZone strings that specify which availability zones the replication group's nodes are to be in. The nummber of PreferredAvailabilityZone values must equal the value of NewReplicaCount plus 1 to account for the primary node. If this member of ReplicaConfiguration is omitted, ElastiCache for Redis selects the availability zone for each of the replicas.

configureShard_preferredOutpostArns :: Lens' ConfigureShard (Maybe [Text])

The outpost ARNs in which the cache cluster is created.

configureShard_nodeGroupId :: Lens' ConfigureShard Text

The 4-digit id for the node group you are configuring. For Redis (cluster mode disabled) replication groups, the node group id is always 0001. To find a Redis (cluster mode enabled)'s node group's (shard's) id, see Finding a Shard's Id.

configureShard_newReplicaCount :: Lens' ConfigureShard Int

The number of replicas you want in this node group at the end of this operation. The maximum value for NewReplicaCount is 5. The minimum value depends upon the type of Redis replication group you are working with.

The minimum number of replicas in a shard or replication group is:

Redis (cluster mode disabled)

If Multi-AZ: 1

If Multi-AZ: 0

Redis (cluster mode enabled): 0 (though you will not be able to failover to a replica if your primary node fails)