Network.AWS.ElastiCache.Types.Event

Description

Documentation

data Event

Represents a single occurrence of something interesting within the system. Some examples of events are creating a cluster, adding or removing a cache node, or rebooting a node.

See: newEvent smart constructor.

Constructors

Event'

Fields

message :: Maybe Text

The text of the event.

date :: Maybe ISO8601

The date and time when the event occurred.

sourceIdentifier :: Maybe Text

The identifier for the source of the event. For example, if the event occurred at the cluster level, the identifier would be the name of the cluster.

sourceType :: Maybe SourceType

Specifies the origin of this event - a cluster, a parameter group, a security group, etc.

newEvent :: Event

Create a value of Event with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:message:Event', event_message - The text of the event.

$sel:date:Event', event_date - The date and time when the event occurred.

$sel:sourceIdentifier:Event', event_sourceIdentifier - The identifier for the source of the event. For example, if the event occurred at the cluster level, the identifier would be the name of the cluster.

$sel:sourceType:Event', event_sourceType - Specifies the origin of this event - a cluster, a parameter group, a security group, etc.

event_message :: Lens' Event (Maybe Text)

The text of the event.

event_date :: Lens' Event (Maybe UTCTime)

The date and time when the event occurred.

event_sourceIdentifier :: Lens' Event (Maybe Text)

The identifier for the source of the event. For example, if the event occurred at the cluster level, the identifier would be the name of the cluster.

event_sourceType :: Lens' Event (Maybe SourceType)

Specifies the origin of this event - a cluster, a parameter group, a security group, etc.