Network.AWS.ElastiCache.Types.LogDeliveryConfiguration

Description

Documentation

data LogDeliveryConfiguration

Returns the destination, format and type of the logs.

See: newLogDeliveryConfiguration smart constructor.

Constructors

LogDeliveryConfiguration'

Fields

status :: Maybe LogDeliveryConfigurationStatus

Returns the log delivery configuration status. Values are one of enabling | disabling | modifying | active | error

destinationType :: Maybe DestinationType

Returns the destination type, either cloudwatch-logs or kinesis-firehose.

logType :: Maybe LogType

Refers to slow-log.

message :: Maybe Text

Returns an error message for the log delivery configuration.

logFormat :: Maybe LogFormat

Returns the log format, either JSON or TEXT.

destinationDetails :: Maybe DestinationDetails

Configuration details of either a CloudWatch Logs destination or Kinesis Data Firehose destination.

newLogDeliveryConfiguration :: LogDeliveryConfiguration

Create a value of LogDeliveryConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:LogDeliveryConfiguration', logDeliveryConfiguration_status - Returns the log delivery configuration status. Values are one of enabling | disabling | modifying | active | error

$sel:destinationType:LogDeliveryConfiguration', logDeliveryConfiguration_destinationType - Returns the destination type, either cloudwatch-logs or kinesis-firehose.

$sel:logType:LogDeliveryConfiguration', logDeliveryConfiguration_logType - Refers to slow-log.

$sel:message:LogDeliveryConfiguration', logDeliveryConfiguration_message - Returns an error message for the log delivery configuration.

$sel:logFormat:LogDeliveryConfiguration', logDeliveryConfiguration_logFormat - Returns the log format, either JSON or TEXT.

$sel:destinationDetails:LogDeliveryConfiguration', logDeliveryConfiguration_destinationDetails - Configuration details of either a CloudWatch Logs destination or Kinesis Data Firehose destination.

logDeliveryConfiguration_status :: Lens' LogDeliveryConfiguration (Maybe LogDeliveryConfigurationStatus)

Returns the log delivery configuration status. Values are one of enabling | disabling | modifying | active | error

logDeliveryConfiguration_destinationType :: Lens' LogDeliveryConfiguration (Maybe DestinationType)

Returns the destination type, either cloudwatch-logs or kinesis-firehose.

logDeliveryConfiguration_logType :: Lens' LogDeliveryConfiguration (Maybe LogType)

Refers to slow-log.

logDeliveryConfiguration_message :: Lens' LogDeliveryConfiguration (Maybe Text)

Returns an error message for the log delivery configuration.

logDeliveryConfiguration_logFormat :: Lens' LogDeliveryConfiguration (Maybe LogFormat)

Returns the log format, either JSON or TEXT.

logDeliveryConfiguration_destinationDetails :: Lens' LogDeliveryConfiguration (Maybe DestinationDetails)

Configuration details of either a CloudWatch Logs destination or Kinesis Data Firehose destination.