Network.AWS.ElastiCache.Types.NodeGroup

Description

Documentation

data NodeGroup

Represents a collection of cache nodes in a replication group. One node in the node group is the read/write primary node. All the other nodes are read-only Replica nodes.

See: newNodeGroup smart constructor.

Constructors

NodeGroup'

Fields

status :: Maybe Text

The current state of this replication group - creating, available, modifying, deleting.

readerEndpoint :: Maybe Endpoint

The endpoint of the replica nodes in this node group (shard).

nodeGroupId :: Maybe Text

The identifier for the node group (shard). A Redis (cluster mode disabled) replication group contains only 1 node group; therefore, the node group ID is 0001. A Redis (cluster mode enabled) replication group contains 1 to 90 node groups numbered 0001 to 0090. Optionally, the user can provide the id for a node group.

slots :: Maybe Text

The keyspace for this node group (shard).

primaryEndpoint :: Maybe Endpoint

The endpoint of the primary node in this node group (shard).

nodeGroupMembers :: Maybe [NodeGroupMember]

A list containing information about individual nodes within the node group (shard).

newNodeGroup :: NodeGroup

Create a value of NodeGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:NodeGroup', nodeGroup_status - The current state of this replication group - creating, available, modifying, deleting.

$sel:readerEndpoint:NodeGroup', nodeGroup_readerEndpoint - The endpoint of the replica nodes in this node group (shard).

$sel:nodeGroupId:NodeGroup', nodeGroup_nodeGroupId - The identifier for the node group (shard). A Redis (cluster mode disabled) replication group contains only 1 node group; therefore, the node group ID is 0001. A Redis (cluster mode enabled) replication group contains 1 to 90 node groups numbered 0001 to 0090. Optionally, the user can provide the id for a node group.

$sel:slots:NodeGroup', nodeGroup_slots - The keyspace for this node group (shard).

$sel:primaryEndpoint:NodeGroup', nodeGroup_primaryEndpoint - The endpoint of the primary node in this node group (shard).

$sel:nodeGroupMembers:NodeGroup', nodeGroup_nodeGroupMembers - A list containing information about individual nodes within the node group (shard).

nodeGroup_status :: Lens' NodeGroup (Maybe Text)

The current state of this replication group - creating, available, modifying, deleting.

nodeGroup_readerEndpoint :: Lens' NodeGroup (Maybe Endpoint)

The endpoint of the replica nodes in this node group (shard).

nodeGroup_nodeGroupId :: Lens' NodeGroup (Maybe Text)

The identifier for the node group (shard). A Redis (cluster mode disabled) replication group contains only 1 node group; therefore, the node group ID is 0001. A Redis (cluster mode enabled) replication group contains 1 to 90 node groups numbered 0001 to 0090. Optionally, the user can provide the id for a node group.

nodeGroup_slots :: Lens' NodeGroup (Maybe Text)

The keyspace for this node group (shard).

nodeGroup_primaryEndpoint :: Lens' NodeGroup (Maybe Endpoint)

The endpoint of the primary node in this node group (shard).

nodeGroup_nodeGroupMembers :: Lens' NodeGroup (Maybe [NodeGroupMember])

A list containing information about individual nodes within the node group (shard).