Network.AWS.ElastiCache.Types.NodeGroupMemberUpdateStatus

Description

Documentation

data NodeGroupMemberUpdateStatus

The status of the service update on the node group member

See: newNodeGroupMemberUpdateStatus smart constructor.

Constructors

NodeGroupMemberUpdateStatus'

Fields

cacheClusterId :: Maybe Text

The cache cluster ID

nodeUpdateStatusModifiedDate :: Maybe ISO8601

The date when the NodeUpdateStatus was last modified

nodeUpdateInitiatedBy :: Maybe NodeUpdateInitiatedBy

Reflects whether the update was initiated by the customer or automatically applied

nodeUpdateStatus :: Maybe NodeUpdateStatus

The update status of the node

nodeUpdateInitiatedDate :: Maybe ISO8601

The date when the update is triggered

cacheNodeId :: Maybe Text

The node ID of the cache cluster

nodeDeletionDate :: Maybe ISO8601

The deletion date of the node

nodeUpdateStartDate :: Maybe ISO8601

The start date of the update for a node

nodeUpdateEndDate :: Maybe ISO8601

The end date of the update for a node

newNodeGroupMemberUpdateStatus :: NodeGroupMemberUpdateStatus

Create a value of NodeGroupMemberUpdateStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cacheClusterId:NodeGroupMemberUpdateStatus', nodeGroupMemberUpdateStatus_cacheClusterId - The cache cluster ID

$sel:nodeUpdateStatusModifiedDate:NodeGroupMemberUpdateStatus', nodeGroupMemberUpdateStatus_nodeUpdateStatusModifiedDate - The date when the NodeUpdateStatus was last modified

$sel:nodeUpdateInitiatedBy:NodeGroupMemberUpdateStatus', nodeGroupMemberUpdateStatus_nodeUpdateInitiatedBy - Reflects whether the update was initiated by the customer or automatically applied

$sel:nodeUpdateStatus:NodeGroupMemberUpdateStatus', nodeGroupMemberUpdateStatus_nodeUpdateStatus - The update status of the node

$sel:nodeUpdateInitiatedDate:NodeGroupMemberUpdateStatus', nodeGroupMemberUpdateStatus_nodeUpdateInitiatedDate - The date when the update is triggered

$sel:cacheNodeId:NodeGroupMemberUpdateStatus', nodeGroupMemberUpdateStatus_cacheNodeId - The node ID of the cache cluster

$sel:nodeDeletionDate:NodeGroupMemberUpdateStatus', nodeGroupMemberUpdateStatus_nodeDeletionDate - The deletion date of the node

$sel:nodeUpdateStartDate:NodeGroupMemberUpdateStatus', nodeGroupMemberUpdateStatus_nodeUpdateStartDate - The start date of the update for a node

$sel:nodeUpdateEndDate:NodeGroupMemberUpdateStatus', nodeGroupMemberUpdateStatus_nodeUpdateEndDate - The end date of the update for a node

nodeGroupMemberUpdateStatus_cacheClusterId :: Lens' NodeGroupMemberUpdateStatus (Maybe Text)

The cache cluster ID

nodeGroupMemberUpdateStatus_nodeUpdateStatusModifiedDate :: Lens' NodeGroupMemberUpdateStatus (Maybe UTCTime)

The date when the NodeUpdateStatus was last modified

nodeGroupMemberUpdateStatus_nodeUpdateInitiatedBy :: Lens' NodeGroupMemberUpdateStatus (Maybe NodeUpdateInitiatedBy)

Reflects whether the update was initiated by the customer or automatically applied

nodeGroupMemberUpdateStatus_nodeUpdateStatus :: Lens' NodeGroupMemberUpdateStatus (Maybe NodeUpdateStatus)

The update status of the node

nodeGroupMemberUpdateStatus_nodeUpdateInitiatedDate :: Lens' NodeGroupMemberUpdateStatus (Maybe UTCTime)

The date when the update is triggered

nodeGroupMemberUpdateStatus_cacheNodeId :: Lens' NodeGroupMemberUpdateStatus (Maybe Text)

The node ID of the cache cluster

nodeGroupMemberUpdateStatus_nodeDeletionDate :: Lens' NodeGroupMemberUpdateStatus (Maybe UTCTime)

The deletion date of the node

nodeGroupMemberUpdateStatus_nodeUpdateStartDate :: Lens' NodeGroupMemberUpdateStatus (Maybe UTCTime)

The start date of the update for a node

nodeGroupMemberUpdateStatus_nodeUpdateEndDate :: Lens' NodeGroupMemberUpdateStatus (Maybe UTCTime)

The end date of the update for a node