Network.AWS.ElastiCache.Types.NodeSnapshot

Description

Documentation

data NodeSnapshot

Represents an individual cache node in a snapshot of a cluster.

See: newNodeSnapshot smart constructor.

Constructors

NodeSnapshot'

Fields

nodeGroupConfiguration :: Maybe NodeGroupConfiguration

The configuration for the source node group (shard).

cacheNodeCreateTime :: Maybe ISO8601

The date and time when the cache node was created in the source cluster.

cacheClusterId :: Maybe Text

A unique identifier for the source cluster.

cacheNodeId :: Maybe Text

The cache node identifier for the node in the source cluster.

nodeGroupId :: Maybe Text

A unique identifier for the source node group (shard).

snapshotCreateTime :: Maybe ISO8601

The date and time when the source node's metadata and cache data set was obtained for the snapshot.

cacheSize :: Maybe Text

The size of the cache on the source cache node.

newNodeSnapshot :: NodeSnapshot

Create a value of NodeSnapshot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nodeGroupConfiguration:NodeSnapshot', nodeSnapshot_nodeGroupConfiguration - The configuration for the source node group (shard).

$sel:cacheNodeCreateTime:NodeSnapshot', nodeSnapshot_cacheNodeCreateTime - The date and time when the cache node was created in the source cluster.

$sel:cacheClusterId:NodeSnapshot', nodeSnapshot_cacheClusterId - A unique identifier for the source cluster.

$sel:cacheNodeId:NodeSnapshot', nodeSnapshot_cacheNodeId - The cache node identifier for the node in the source cluster.

$sel:nodeGroupId:NodeSnapshot', nodeSnapshot_nodeGroupId - A unique identifier for the source node group (shard).

$sel:snapshotCreateTime:NodeSnapshot', nodeSnapshot_snapshotCreateTime - The date and time when the source node's metadata and cache data set was obtained for the snapshot.

$sel:cacheSize:NodeSnapshot', nodeSnapshot_cacheSize - The size of the cache on the source cache node.

nodeSnapshot_nodeGroupConfiguration :: Lens' NodeSnapshot (Maybe NodeGroupConfiguration)

The configuration for the source node group (shard).

nodeSnapshot_cacheNodeCreateTime :: Lens' NodeSnapshot (Maybe UTCTime)

The date and time when the cache node was created in the source cluster.

nodeSnapshot_cacheClusterId :: Lens' NodeSnapshot (Maybe Text)

A unique identifier for the source cluster.

nodeSnapshot_cacheNodeId :: Lens' NodeSnapshot (Maybe Text)

The cache node identifier for the node in the source cluster.

nodeSnapshot_nodeGroupId :: Lens' NodeSnapshot (Maybe Text)

A unique identifier for the source node group (shard).

nodeSnapshot_snapshotCreateTime :: Lens' NodeSnapshot (Maybe UTCTime)

The date and time when the source node's metadata and cache data set was obtained for the snapshot.

nodeSnapshot_cacheSize :: Lens' NodeSnapshot (Maybe Text)

The size of the cache on the source cache node.