Network.AWS.ElastiCache.Types.RegionalConfiguration

Description

Documentation

data RegionalConfiguration

A list of the replication groups

See: newRegionalConfiguration smart constructor.

Constructors

RegionalConfiguration'

Fields

replicationGroupId :: Text

The name of the secondary cluster

replicationGroupRegion :: Text

The Amazon region where the cluster is stored

reshardingConfiguration :: [ReshardingConfiguration]

A list of PreferredAvailabilityZones objects that specifies the configuration of a node group in the resharded cluster.

newRegionalConfiguration

Arguments

:: Text
$sel:replicationGroupId:RegionalConfiguration'

-> Text
$sel:replicationGroupRegion:RegionalConfiguration'

-> RegionalConfiguration

Create a value of RegionalConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:replicationGroupId:RegionalConfiguration', regionalConfiguration_replicationGroupId - The name of the secondary cluster

$sel:replicationGroupRegion:RegionalConfiguration', regionalConfiguration_replicationGroupRegion - The Amazon region where the cluster is stored

$sel:reshardingConfiguration:RegionalConfiguration', regionalConfiguration_reshardingConfiguration - A list of PreferredAvailabilityZones objects that specifies the configuration of a node group in the resharded cluster.

regionalConfiguration_replicationGroupId :: Lens' RegionalConfiguration Text

The name of the secondary cluster

regionalConfiguration_replicationGroupRegion :: Lens' RegionalConfiguration Text

The Amazon region where the cluster is stored

regionalConfiguration_reshardingConfiguration :: Lens' RegionalConfiguration [ReshardingConfiguration]

A list of PreferredAvailabilityZones objects that specifies the configuration of a node group in the resharded cluster.