Network.AWS.ElastiCache.Types.ReplicationGroup

Description

Documentation

data ReplicationGroup

Contains all of the attributes of a specific Redis replication group.

See: newReplicationGroup smart constructor.

Constructors

ReplicationGroup'

Fields

clusterEnabled :: Maybe Bool

A flag indicating whether or not this replication group is cluster enabled; i.e., whether its data can be partitioned across multiple shards (API/CLI: node groups).

Valid values: true | false

status :: Maybe Text

The current state of this replication group - creating, available, modifying, deleting, create-failed, snapshotting.

nodeGroups :: Maybe [NodeGroup]

A list of node groups in this replication group. For Redis (cluster mode disabled) replication groups, this is a single-element list. For Redis (cluster mode enabled) replication groups, the list contains an entry for each node group (shard).

automaticFailover :: Maybe AutomaticFailoverStatus

Indicates the status of automatic failover for this Redis replication group.

memberClustersOutpostArns :: Maybe [Text]

The outpost ARNs of the replication group's member clusters.

memberClusters :: Maybe [Text]

The names of all the cache clusters that are part of this replication group.

replicationGroupId :: Maybe Text

The identifier for the replication group.

globalReplicationGroupInfo :: Maybe GlobalReplicationGroupInfo

The name of the Global datastore and role of this replication group in the Global datastore.

userGroupIds :: Maybe [Text]

The ID of the user group associated to the replication group.

snapshotWindow :: Maybe Text

The daily time range (in UTC) during which ElastiCache begins taking a daily snapshot of your node group (shard).

Example: 05:00-09:00

If you do not specify this parameter, ElastiCache automatically chooses an appropriate time range.

This parameter is only valid if the Engine parameter is redis.

snapshotRetentionLimit :: Maybe Int

The number of days for which ElastiCache retains automatic cluster snapshots before deleting them. For example, if you set SnapshotRetentionLimit to 5, a snapshot that was taken today is retained for 5 days before being deleted.

If the value of SnapshotRetentionLimit is set to zero (0), backups are turned off.

arn :: Maybe Text

The ARN (Amazon Resource Name) of the replication group.

multiAZ :: Maybe MultiAZStatus

A flag indicating if you have Multi-AZ enabled to enhance fault tolerance. For more information, see Minimizing Downtime: Multi-AZ

replicationGroupCreateTime :: Maybe ISO8601

The date and time when the cluster was created.

kmsKeyId :: Maybe Text

The ID of the KMS key used to encrypt the disk in the cluster.

snapshottingClusterId :: Maybe Text

The cluster ID that is used as the daily snapshot source for the replication group.

atRestEncryptionEnabled :: Maybe Bool

A flag that enables encryption at-rest when set to true.

You cannot modify the value of AtRestEncryptionEnabled after the cluster is created. To enable encryption at-rest on a cluster you must set AtRestEncryptionEnabled to true when you create a cluster.

Required: Only available when creating a replication group in an Amazon VPC using redis version 3.2.6, 4.x or later.

Default: false

authTokenLastModifiedDate :: Maybe ISO8601

The date the auth token was last modified

cacheNodeType :: Maybe Text

The name of the compute and memory capacity node type for each node in the replication group.

description :: Maybe Text

The user supplied description of the replication group.

pendingModifiedValues :: Maybe ReplicationGroupPendingModifiedValues

A group of settings to be applied to the replication group, either immediately or during the next maintenance window.

authTokenEnabled :: Maybe Bool

A flag that enables using an AuthToken (password) when issuing Redis commands.

Default: false

configurationEndpoint :: Maybe Endpoint

The configuration endpoint for this replication group. Use the configuration endpoint to connect to this replication group.

logDeliveryConfigurations :: Maybe [LogDeliveryConfiguration]

Returns the destination, format and type of the logs.

transitEncryptionEnabled :: Maybe Bool

A flag that enables in-transit encryption when set to true.

You cannot modify the value of TransitEncryptionEnabled after the cluster is created. To enable in-transit encryption on a cluster you must set TransitEncryptionEnabled to true when you create a cluster.

Required: Only available when creating a replication group in an Amazon VPC using redis version 3.2.6, 4.x or later.

Default: false

newReplicationGroup :: ReplicationGroup

Create a value of ReplicationGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clusterEnabled:ReplicationGroup', replicationGroup_clusterEnabled - A flag indicating whether or not this replication group is cluster enabled; i.e., whether its data can be partitioned across multiple shards (API/CLI: node groups).

Valid values: true | false

$sel:status:ReplicationGroup', replicationGroup_status - The current state of this replication group - creating, available, modifying, deleting, create-failed, snapshotting.

$sel:nodeGroups:ReplicationGroup', replicationGroup_nodeGroups - A list of node groups in this replication group. For Redis (cluster mode disabled) replication groups, this is a single-element list. For Redis (cluster mode enabled) replication groups, the list contains an entry for each node group (shard).

$sel:automaticFailover:ReplicationGroup', replicationGroup_automaticFailover - Indicates the status of automatic failover for this Redis replication group.

$sel:memberClustersOutpostArns:ReplicationGroup', replicationGroup_memberClustersOutpostArns - The outpost ARNs of the replication group's member clusters.

$sel:memberClusters:ReplicationGroup', replicationGroup_memberClusters - The names of all the cache clusters that are part of this replication group.

$sel:replicationGroupId:ReplicationGroup', replicationGroup_replicationGroupId - The identifier for the replication group.

$sel:globalReplicationGroupInfo:ReplicationGroup', replicationGroup_globalReplicationGroupInfo - The name of the Global datastore and role of this replication group in the Global datastore.

$sel:userGroupIds:ReplicationGroup', replicationGroup_userGroupIds - The ID of the user group associated to the replication group.

$sel:snapshotWindow:ReplicationGroup', replicationGroup_snapshotWindow - The daily time range (in UTC) during which ElastiCache begins taking a daily snapshot of your node group (shard).

Example: 05:00-09:00

If you do not specify this parameter, ElastiCache automatically chooses an appropriate time range.

This parameter is only valid if the Engine parameter is redis.

$sel:snapshotRetentionLimit:ReplicationGroup', replicationGroup_snapshotRetentionLimit - The number of days for which ElastiCache retains automatic cluster snapshots before deleting them. For example, if you set SnapshotRetentionLimit to 5, a snapshot that was taken today is retained for 5 days before being deleted.

If the value of SnapshotRetentionLimit is set to zero (0), backups are turned off.

$sel:arn:ReplicationGroup', replicationGroup_arn - The ARN (Amazon Resource Name) of the replication group.

$sel:multiAZ:ReplicationGroup', replicationGroup_multiAZ - A flag indicating if you have Multi-AZ enabled to enhance fault tolerance. For more information, see Minimizing Downtime: Multi-AZ

$sel:replicationGroupCreateTime:ReplicationGroup', replicationGroup_replicationGroupCreateTime - The date and time when the cluster was created.

$sel:kmsKeyId:ReplicationGroup', replicationGroup_kmsKeyId - The ID of the KMS key used to encrypt the disk in the cluster.

$sel:snapshottingClusterId:ReplicationGroup', replicationGroup_snapshottingClusterId - The cluster ID that is used as the daily snapshot source for the replication group.

$sel:atRestEncryptionEnabled:ReplicationGroup', replicationGroup_atRestEncryptionEnabled - A flag that enables encryption at-rest when set to true.

You cannot modify the value of AtRestEncryptionEnabled after the cluster is created. To enable encryption at-rest on a cluster you must set AtRestEncryptionEnabled to true when you create a cluster.

Required: Only available when creating a replication group in an Amazon VPC using redis version 3.2.6, 4.x or later.

Default: false

$sel:authTokenLastModifiedDate:ReplicationGroup', replicationGroup_authTokenLastModifiedDate - The date the auth token was last modified

$sel:cacheNodeType:ReplicationGroup', replicationGroup_cacheNodeType - The name of the compute and memory capacity node type for each node in the replication group.

$sel:description:ReplicationGroup', replicationGroup_description - The user supplied description of the replication group.

$sel:pendingModifiedValues:ReplicationGroup', replicationGroup_pendingModifiedValues - A group of settings to be applied to the replication group, either immediately or during the next maintenance window.

$sel:authTokenEnabled:ReplicationGroup', replicationGroup_authTokenEnabled - A flag that enables using an AuthToken (password) when issuing Redis commands.

Default: false

$sel:configurationEndpoint:ReplicationGroup', replicationGroup_configurationEndpoint - The configuration endpoint for this replication group. Use the configuration endpoint to connect to this replication group.

$sel:logDeliveryConfigurations:ReplicationGroup', replicationGroup_logDeliveryConfigurations - Returns the destination, format and type of the logs.

$sel:transitEncryptionEnabled:ReplicationGroup', replicationGroup_transitEncryptionEnabled - A flag that enables in-transit encryption when set to true.

You cannot modify the value of TransitEncryptionEnabled after the cluster is created. To enable in-transit encryption on a cluster you must set TransitEncryptionEnabled to true when you create a cluster.

Required: Only available when creating a replication group in an Amazon VPC using redis version 3.2.6, 4.x or later.

Default: false

replicationGroup_clusterEnabled :: Lens' ReplicationGroup (Maybe Bool)

A flag indicating whether or not this replication group is cluster enabled; i.e., whether its data can be partitioned across multiple shards (API/CLI: node groups).

Valid values: true | false

replicationGroup_status :: Lens' ReplicationGroup (Maybe Text)

The current state of this replication group - creating, available, modifying, deleting, create-failed, snapshotting.

replicationGroup_nodeGroups :: Lens' ReplicationGroup (Maybe [NodeGroup])

A list of node groups in this replication group. For Redis (cluster mode disabled) replication groups, this is a single-element list. For Redis (cluster mode enabled) replication groups, the list contains an entry for each node group (shard).

replicationGroup_automaticFailover :: Lens' ReplicationGroup (Maybe AutomaticFailoverStatus)

Indicates the status of automatic failover for this Redis replication group.

replicationGroup_memberClustersOutpostArns :: Lens' ReplicationGroup (Maybe [Text])

The outpost ARNs of the replication group's member clusters.

replicationGroup_memberClusters :: Lens' ReplicationGroup (Maybe [Text])

The names of all the cache clusters that are part of this replication group.

replicationGroup_replicationGroupId :: Lens' ReplicationGroup (Maybe Text)

The identifier for the replication group.

replicationGroup_globalReplicationGroupInfo :: Lens' ReplicationGroup (Maybe GlobalReplicationGroupInfo)

The name of the Global datastore and role of this replication group in the Global datastore.

replicationGroup_userGroupIds :: Lens' ReplicationGroup (Maybe [Text])

The ID of the user group associated to the replication group.

replicationGroup_snapshotWindow :: Lens' ReplicationGroup (Maybe Text)

The daily time range (in UTC) during which ElastiCache begins taking a daily snapshot of your node group (shard).

Example: 05:00-09:00

If you do not specify this parameter, ElastiCache automatically chooses an appropriate time range.

This parameter is only valid if the Engine parameter is redis.

replicationGroup_snapshotRetentionLimit :: Lens' ReplicationGroup (Maybe Int)

The number of days for which ElastiCache retains automatic cluster snapshots before deleting them. For example, if you set SnapshotRetentionLimit to 5, a snapshot that was taken today is retained for 5 days before being deleted.

If the value of SnapshotRetentionLimit is set to zero (0), backups are turned off.

replicationGroup_arn :: Lens' ReplicationGroup (Maybe Text)

The ARN (Amazon Resource Name) of the replication group.

replicationGroup_multiAZ :: Lens' ReplicationGroup (Maybe MultiAZStatus)

A flag indicating if you have Multi-AZ enabled to enhance fault tolerance. For more information, see Minimizing Downtime: Multi-AZ

replicationGroup_replicationGroupCreateTime :: Lens' ReplicationGroup (Maybe UTCTime)

The date and time when the cluster was created.

replicationGroup_kmsKeyId :: Lens' ReplicationGroup (Maybe Text)

The ID of the KMS key used to encrypt the disk in the cluster.

replicationGroup_snapshottingClusterId :: Lens' ReplicationGroup (Maybe Text)

The cluster ID that is used as the daily snapshot source for the replication group.

replicationGroup_atRestEncryptionEnabled :: Lens' ReplicationGroup (Maybe Bool)

A flag that enables encryption at-rest when set to true.

You cannot modify the value of AtRestEncryptionEnabled after the cluster is created. To enable encryption at-rest on a cluster you must set AtRestEncryptionEnabled to true when you create a cluster.

Required: Only available when creating a replication group in an Amazon VPC using redis version 3.2.6, 4.x or later.

Default: false

replicationGroup_authTokenLastModifiedDate :: Lens' ReplicationGroup (Maybe UTCTime)

The date the auth token was last modified

replicationGroup_cacheNodeType :: Lens' ReplicationGroup (Maybe Text)

The name of the compute and memory capacity node type for each node in the replication group.

replicationGroup_description :: Lens' ReplicationGroup (Maybe Text)

The user supplied description of the replication group.

replicationGroup_pendingModifiedValues :: Lens' ReplicationGroup (Maybe ReplicationGroupPendingModifiedValues)

A group of settings to be applied to the replication group, either immediately or during the next maintenance window.

replicationGroup_authTokenEnabled :: Lens' ReplicationGroup (Maybe Bool)

A flag that enables using an AuthToken (password) when issuing Redis commands.

Default: false

replicationGroup_configurationEndpoint :: Lens' ReplicationGroup (Maybe Endpoint)

The configuration endpoint for this replication group. Use the configuration endpoint to connect to this replication group.

replicationGroup_logDeliveryConfigurations :: Lens' ReplicationGroup (Maybe [LogDeliveryConfiguration])

Returns the destination, format and type of the logs.

replicationGroup_transitEncryptionEnabled :: Lens' ReplicationGroup (Maybe Bool)

A flag that enables in-transit encryption when set to true.

You cannot modify the value of TransitEncryptionEnabled after the cluster is created. To enable in-transit encryption on a cluster you must set TransitEncryptionEnabled to true when you create a cluster.

Required: Only available when creating a replication group in an Amazon VPC using redis version 3.2.6, 4.x or later.

Default: false