Network.AWS.ElastiCache.Types.ReplicationGroupPendingModifiedValues

Description

Documentation

data ReplicationGroupPendingModifiedValues

The settings to be applied to the Redis replication group, either immediately or during the next maintenance window.

See: newReplicationGroupPendingModifiedValues smart constructor.

Constructors

ReplicationGroupPendingModifiedValues'

Fields

authTokenStatus :: Maybe AuthTokenUpdateStatus

The auth token status

userGroups :: Maybe UserGroupsUpdateStatus

The user group being modified.

logDeliveryConfigurations :: Maybe [PendingLogDeliveryConfiguration]

The log delivery configurations being modified

resharding :: Maybe ReshardingStatus

The status of an online resharding operation.

primaryClusterId :: Maybe Text

The primary cluster ID that is applied immediately (if --apply-immediately was specified), or during the next maintenance window.

automaticFailoverStatus :: Maybe PendingAutomaticFailoverStatus

Indicates the status of automatic failover for this Redis replication group.

newReplicationGroupPendingModifiedValues :: ReplicationGroupPendingModifiedValues

Create a value of ReplicationGroupPendingModifiedValues with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:authTokenStatus:ReplicationGroupPendingModifiedValues', replicationGroupPendingModifiedValues_authTokenStatus - The auth token status

$sel:userGroups:ReplicationGroupPendingModifiedValues', replicationGroupPendingModifiedValues_userGroups - The user group being modified.

$sel:logDeliveryConfigurations:ReplicationGroupPendingModifiedValues', replicationGroupPendingModifiedValues_logDeliveryConfigurations - The log delivery configurations being modified

$sel:resharding:ReplicationGroupPendingModifiedValues', replicationGroupPendingModifiedValues_resharding - The status of an online resharding operation.

$sel:primaryClusterId:ReplicationGroupPendingModifiedValues', replicationGroupPendingModifiedValues_primaryClusterId - The primary cluster ID that is applied immediately (if --apply-immediately was specified), or during the next maintenance window.

$sel:automaticFailoverStatus:ReplicationGroupPendingModifiedValues', replicationGroupPendingModifiedValues_automaticFailoverStatus - Indicates the status of automatic failover for this Redis replication group.

replicationGroupPendingModifiedValues_authTokenStatus :: Lens' ReplicationGroupPendingModifiedValues (Maybe AuthTokenUpdateStatus)

The auth token status

replicationGroupPendingModifiedValues_userGroups :: Lens' ReplicationGroupPendingModifiedValues (Maybe UserGroupsUpdateStatus)

The user group being modified.

replicationGroupPendingModifiedValues_logDeliveryConfigurations :: Lens' ReplicationGroupPendingModifiedValues (Maybe [PendingLogDeliveryConfiguration])

The log delivery configurations being modified

replicationGroupPendingModifiedValues_resharding :: Lens' ReplicationGroupPendingModifiedValues (Maybe ReshardingStatus)

The status of an online resharding operation.

replicationGroupPendingModifiedValues_primaryClusterId :: Lens' ReplicationGroupPendingModifiedValues (Maybe Text)

The primary cluster ID that is applied immediately (if --apply-immediately was specified), or during the next maintenance window.

replicationGroupPendingModifiedValues_automaticFailoverStatus :: Lens' ReplicationGroupPendingModifiedValues (Maybe PendingAutomaticFailoverStatus)

Indicates the status of automatic failover for this Redis replication group.