Network.AWS.ElastiCache.Types.ReservedCacheNode

Description

Documentation

data ReservedCacheNode

Represents the output of a PurchaseReservedCacheNodesOffering operation.

See: newReservedCacheNode smart constructor.

Constructors

ReservedCacheNode'

Fields

duration :: Maybe Int

The duration of the reservation in seconds.

reservedCacheNodesOfferingId :: Maybe Text

The offering identifier.

startTime :: Maybe ISO8601

The time the reservation started.

cacheNodeCount :: Maybe Int

The number of cache nodes that have been reserved.

state :: Maybe Text

The state of the reserved cache node.

cacheNodeType :: Maybe Text

The cache node type for the reserved cache nodes.

The following node types are supported by ElastiCache. Generally speaking, the current generation types provide more memory and computational power at lower cost when compared to their equivalent previous generation counterparts.

General purpose:

Current generation:

M6g node types (available only for Redis engine version 5.0.6 onward and for Memcached engine version 1.5.16 onward).

cache.m6g.large, cache.m6g.xlarge, cache.m6g.2xlarge, cache.m6g.4xlarge, cache.m6g.8xlarge, cache.m6g.12xlarge, cache.m6g.16xlarge

For region availability, see Supported Node Types

M5 node types: cache.m5.large, cache.m5.xlarge, cache.m5.2xlarge, cache.m5.4xlarge, cache.m5.12xlarge, cache.m5.24xlarge

M4 node types: cache.m4.large, cache.m4.xlarge, cache.m4.2xlarge, cache.m4.4xlarge, cache.m4.10xlarge

T3 node types: cache.t3.micro, cache.t3.small, cache.t3.medium

T2 node types: cache.t2.micro, cache.t2.small, cache.t2.medium

Previous generation: (not recommended)

T1 node types: cache.t1.micro

M1 node types: cache.m1.small, cache.m1.medium, cache.m1.large, cache.m1.xlarge

M3 node types: cache.m3.medium, cache.m3.large, cache.m3.xlarge, cache.m3.2xlarge

Compute optimized:

Previous generation: (not recommended)

C1 node types: cache.c1.xlarge

Memory optimized:

Current generation:

R6g node types (available only for Redis engine version 5.0.6 onward and for Memcached engine version 1.5.16 onward).

cache.r6g.large, cache.r6g.xlarge, cache.r6g.2xlarge, cache.r6g.4xlarge, cache.r6g.8xlarge, cache.r6g.12xlarge, cache.r6g.16xlarge

For region availability, see Supported Node Types

R5 node types: cache.r5.large, cache.r5.xlarge, cache.r5.2xlarge, cache.r5.4xlarge, cache.r5.12xlarge, cache.r5.24xlarge

R4 node types: cache.r4.large, cache.r4.xlarge, cache.r4.2xlarge, cache.r4.4xlarge, cache.r4.8xlarge, cache.r4.16xlarge

Previous generation: (not recommended)

M2 node types: cache.m2.xlarge, cache.m2.2xlarge, cache.m2.4xlarge

R3 node types: cache.r3.large, cache.r3.xlarge, cache.r3.2xlarge, cache.r3.4xlarge, cache.r3.8xlarge

Additional node type info

All current generation instance types are created in Amazon VPC by default.

Redis append-only files (AOF) are not supported for T1 or T2 instances.

Redis Multi-AZ with automatic failover is not supported on T1 instances.

Redis configuration variables appendonly and appendfsync are not supported on Redis version 2.8.22 and later.

fixedPrice :: Maybe Double

The fixed price charged for this reserved cache node.

usagePrice :: Maybe Double

The hourly price charged for this reserved cache node.

offeringType :: Maybe Text

The offering type of this reserved cache node.

recurringCharges :: Maybe [RecurringCharge]

The recurring price charged to run this reserved cache node.

reservedCacheNodeId :: Maybe Text

The unique identifier for the reservation.

productDescription :: Maybe Text

The description of the reserved cache node.

reservationARN :: Maybe Text

The Amazon Resource Name (ARN) of the reserved cache node.

Example: arn:aws:elasticache:us-east-1:123456789012:reserved-instance:ri-2017-03-27-08-33-25-582

newReservedCacheNode :: ReservedCacheNode

Create a value of ReservedCacheNode with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:duration:ReservedCacheNode', reservedCacheNode_duration - The duration of the reservation in seconds.

$sel:reservedCacheNodesOfferingId:ReservedCacheNode', reservedCacheNode_reservedCacheNodesOfferingId - The offering identifier.

$sel:startTime:ReservedCacheNode', reservedCacheNode_startTime - The time the reservation started.

$sel:cacheNodeCount:ReservedCacheNode', reservedCacheNode_cacheNodeCount - The number of cache nodes that have been reserved.

$sel:state:ReservedCacheNode', reservedCacheNode_state - The state of the reserved cache node.

$sel:cacheNodeType:ReservedCacheNode', reservedCacheNode_cacheNodeType - The cache node type for the reserved cache nodes.

The following node types are supported by ElastiCache. Generally speaking, the current generation types provide more memory and computational power at lower cost when compared to their equivalent previous generation counterparts.

General purpose:

Current generation:

M6g node types (available only for Redis engine version 5.0.6 onward and for Memcached engine version 1.5.16 onward).

cache.m6g.large, cache.m6g.xlarge, cache.m6g.2xlarge, cache.m6g.4xlarge, cache.m6g.8xlarge, cache.m6g.12xlarge, cache.m6g.16xlarge

For region availability, see Supported Node Types

M5 node types: cache.m5.large, cache.m5.xlarge, cache.m5.2xlarge, cache.m5.4xlarge, cache.m5.12xlarge, cache.m5.24xlarge

M4 node types: cache.m4.large, cache.m4.xlarge, cache.m4.2xlarge, cache.m4.4xlarge, cache.m4.10xlarge

T3 node types: cache.t3.micro, cache.t3.small, cache.t3.medium

T2 node types: cache.t2.micro, cache.t2.small, cache.t2.medium

Previous generation: (not recommended)

T1 node types: cache.t1.micro

M1 node types: cache.m1.small, cache.m1.medium, cache.m1.large, cache.m1.xlarge

M3 node types: cache.m3.medium, cache.m3.large, cache.m3.xlarge, cache.m3.2xlarge

Compute optimized:

Previous generation: (not recommended)

C1 node types: cache.c1.xlarge

Memory optimized:

Current generation:

R6g node types (available only for Redis engine version 5.0.6 onward and for Memcached engine version 1.5.16 onward).

cache.r6g.large, cache.r6g.xlarge, cache.r6g.2xlarge, cache.r6g.4xlarge, cache.r6g.8xlarge, cache.r6g.12xlarge, cache.r6g.16xlarge

For region availability, see Supported Node Types

R5 node types: cache.r5.large, cache.r5.xlarge, cache.r5.2xlarge, cache.r5.4xlarge, cache.r5.12xlarge, cache.r5.24xlarge

R4 node types: cache.r4.large, cache.r4.xlarge, cache.r4.2xlarge, cache.r4.4xlarge, cache.r4.8xlarge, cache.r4.16xlarge

Previous generation: (not recommended)

M2 node types: cache.m2.xlarge, cache.m2.2xlarge, cache.m2.4xlarge

R3 node types: cache.r3.large, cache.r3.xlarge, cache.r3.2xlarge, cache.r3.4xlarge, cache.r3.8xlarge

Additional node type info

All current generation instance types are created in Amazon VPC by default.

Redis append-only files (AOF) are not supported for T1 or T2 instances.

Redis Multi-AZ with automatic failover is not supported on T1 instances.

Redis configuration variables appendonly and appendfsync are not supported on Redis version 2.8.22 and later.

$sel:fixedPrice:ReservedCacheNode', reservedCacheNode_fixedPrice - The fixed price charged for this reserved cache node.

$sel:usagePrice:ReservedCacheNode', reservedCacheNode_usagePrice - The hourly price charged for this reserved cache node.

$sel:offeringType:ReservedCacheNode', reservedCacheNode_offeringType - The offering type of this reserved cache node.

$sel:recurringCharges:ReservedCacheNode', reservedCacheNode_recurringCharges - The recurring price charged to run this reserved cache node.

$sel:reservedCacheNodeId:ReservedCacheNode', reservedCacheNode_reservedCacheNodeId - The unique identifier for the reservation.

$sel:productDescription:ReservedCacheNode', reservedCacheNode_productDescription - The description of the reserved cache node.

$sel:reservationARN:ReservedCacheNode', reservedCacheNode_reservationARN - The Amazon Resource Name (ARN) of the reserved cache node.

Example: arn:aws:elasticache:us-east-1:123456789012:reserved-instance:ri-2017-03-27-08-33-25-582

reservedCacheNode_duration :: Lens' ReservedCacheNode (Maybe Int)

The duration of the reservation in seconds.

reservedCacheNode_reservedCacheNodesOfferingId :: Lens' ReservedCacheNode (Maybe Text)

The offering identifier.

reservedCacheNode_startTime :: Lens' ReservedCacheNode (Maybe UTCTime)

The time the reservation started.

reservedCacheNode_cacheNodeCount :: Lens' ReservedCacheNode (Maybe Int)

The number of cache nodes that have been reserved.

reservedCacheNode_state :: Lens' ReservedCacheNode (Maybe Text)

The state of the reserved cache node.

reservedCacheNode_cacheNodeType :: Lens' ReservedCacheNode (Maybe Text)

The cache node type for the reserved cache nodes.

The following node types are supported by ElastiCache. Generally speaking, the current generation types provide more memory and computational power at lower cost when compared to their equivalent previous generation counterparts.

General purpose:

Current generation:

M6g node types (available only for Redis engine version 5.0.6 onward and for Memcached engine version 1.5.16 onward).

cache.m6g.large, cache.m6g.xlarge, cache.m6g.2xlarge, cache.m6g.4xlarge, cache.m6g.8xlarge, cache.m6g.12xlarge, cache.m6g.16xlarge

For region availability, see Supported Node Types

M5 node types: cache.m5.large, cache.m5.xlarge, cache.m5.2xlarge, cache.m5.4xlarge, cache.m5.12xlarge, cache.m5.24xlarge

M4 node types: cache.m4.large, cache.m4.xlarge, cache.m4.2xlarge, cache.m4.4xlarge, cache.m4.10xlarge

T3 node types: cache.t3.micro, cache.t3.small, cache.t3.medium

T2 node types: cache.t2.micro, cache.t2.small, cache.t2.medium

Previous generation: (not recommended)

T1 node types: cache.t1.micro

M1 node types: cache.m1.small, cache.m1.medium, cache.m1.large, cache.m1.xlarge

M3 node types: cache.m3.medium, cache.m3.large, cache.m3.xlarge, cache.m3.2xlarge

Compute optimized:

Previous generation: (not recommended)

C1 node types: cache.c1.xlarge

Memory optimized:

Current generation:

R6g node types (available only for Redis engine version 5.0.6 onward and for Memcached engine version 1.5.16 onward).

cache.r6g.large, cache.r6g.xlarge, cache.r6g.2xlarge, cache.r6g.4xlarge, cache.r6g.8xlarge, cache.r6g.12xlarge, cache.r6g.16xlarge

For region availability, see Supported Node Types

R5 node types: cache.r5.large, cache.r5.xlarge, cache.r5.2xlarge, cache.r5.4xlarge, cache.r5.12xlarge, cache.r5.24xlarge

R4 node types: cache.r4.large, cache.r4.xlarge, cache.r4.2xlarge, cache.r4.4xlarge, cache.r4.8xlarge, cache.r4.16xlarge

Previous generation: (not recommended)

M2 node types: cache.m2.xlarge, cache.m2.2xlarge, cache.m2.4xlarge

R3 node types: cache.r3.large, cache.r3.xlarge, cache.r3.2xlarge, cache.r3.4xlarge, cache.r3.8xlarge

Additional node type info

All current generation instance types are created in Amazon VPC by default.

Redis append-only files (AOF) are not supported for T1 or T2 instances.

Redis Multi-AZ with automatic failover is not supported on T1 instances.

Redis configuration variables appendonly and appendfsync are not supported on Redis version 2.8.22 and later.

reservedCacheNode_fixedPrice :: Lens' ReservedCacheNode (Maybe Double)

The fixed price charged for this reserved cache node.

reservedCacheNode_usagePrice :: Lens' ReservedCacheNode (Maybe Double)

The hourly price charged for this reserved cache node.

reservedCacheNode_offeringType :: Lens' ReservedCacheNode (Maybe Text)

The offering type of this reserved cache node.

reservedCacheNode_recurringCharges :: Lens' ReservedCacheNode (Maybe [RecurringCharge])

The recurring price charged to run this reserved cache node.

reservedCacheNode_reservedCacheNodeId :: Lens' ReservedCacheNode (Maybe Text)

The unique identifier for the reservation.

reservedCacheNode_productDescription :: Lens' ReservedCacheNode (Maybe Text)

The description of the reserved cache node.

reservedCacheNode_reservationARN :: Lens' ReservedCacheNode (Maybe Text)

The Amazon Resource Name (ARN) of the reserved cache node.

Example: arn:aws:elasticache:us-east-1:123456789012:reserved-instance:ri-2017-03-27-08-33-25-582