Network.AWS.ElastiCache.Types.ServiceUpdate

Description

Documentation

data ServiceUpdate

An update that you can apply to your Redis clusters.

See: newServiceUpdate smart constructor.

Constructors

ServiceUpdate'

Fields

engineVersion :: Maybe Text

The Elasticache engine version to which the update applies. Either Redis or Memcached engine version

serviceUpdateType :: Maybe ServiceUpdateType

Reflects the nature of the service update

serviceUpdateName :: Maybe Text

The unique ID of the service update

engine :: Maybe Text

The Elasticache engine to which the update applies. Either Redis or Memcached

serviceUpdateReleaseDate :: Maybe ISO8601

The date when the service update is initially available

autoUpdateAfterRecommendedApplyByDate :: Maybe Bool

Indicates whether the service update will be automatically applied once the recommended apply-by date has expired.

serviceUpdateSeverity :: Maybe ServiceUpdateSeverity

The severity of the service update

serviceUpdateEndDate :: Maybe ISO8601

The date after which the service update is no longer available

serviceUpdateDescription :: Maybe Text

Provides details of the service update

serviceUpdateRecommendedApplyByDate :: Maybe ISO8601

The recommendend date to apply the service update in order to ensure compliance. For information on compliance, see Self-Service Security Updates for Compliance.

serviceUpdateStatus :: Maybe ServiceUpdateStatus

The status of the service update

estimatedUpdateTime :: Maybe Text

The estimated length of time the service update will take

newServiceUpdate :: ServiceUpdate

Create a value of ServiceUpdate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:engineVersion:ServiceUpdate', serviceUpdate_engineVersion - The Elasticache engine version to which the update applies. Either Redis or Memcached engine version

$sel:serviceUpdateType:ServiceUpdate', serviceUpdate_serviceUpdateType - Reflects the nature of the service update

$sel:serviceUpdateName:ServiceUpdate', serviceUpdate_serviceUpdateName - The unique ID of the service update

$sel:engine:ServiceUpdate', serviceUpdate_engine - The Elasticache engine to which the update applies. Either Redis or Memcached

$sel:serviceUpdateReleaseDate:ServiceUpdate', serviceUpdate_serviceUpdateReleaseDate - The date when the service update is initially available

$sel:autoUpdateAfterRecommendedApplyByDate:ServiceUpdate', serviceUpdate_autoUpdateAfterRecommendedApplyByDate - Indicates whether the service update will be automatically applied once the recommended apply-by date has expired.

$sel:serviceUpdateSeverity:ServiceUpdate', serviceUpdate_serviceUpdateSeverity - The severity of the service update

$sel:serviceUpdateEndDate:ServiceUpdate', serviceUpdate_serviceUpdateEndDate - The date after which the service update is no longer available

$sel:serviceUpdateDescription:ServiceUpdate', serviceUpdate_serviceUpdateDescription - Provides details of the service update

$sel:serviceUpdateRecommendedApplyByDate:ServiceUpdate', serviceUpdate_serviceUpdateRecommendedApplyByDate - The recommendend date to apply the service update in order to ensure compliance. For information on compliance, see Self-Service Security Updates for Compliance.

$sel:serviceUpdateStatus:ServiceUpdate', serviceUpdate_serviceUpdateStatus - The status of the service update

$sel:estimatedUpdateTime:ServiceUpdate', serviceUpdate_estimatedUpdateTime - The estimated length of time the service update will take

serviceUpdate_engineVersion :: Lens' ServiceUpdate (Maybe Text)

The Elasticache engine version to which the update applies. Either Redis or Memcached engine version

serviceUpdate_serviceUpdateType :: Lens' ServiceUpdate (Maybe ServiceUpdateType)

Reflects the nature of the service update

serviceUpdate_serviceUpdateName :: Lens' ServiceUpdate (Maybe Text)

The unique ID of the service update

serviceUpdate_engine :: Lens' ServiceUpdate (Maybe Text)

The Elasticache engine to which the update applies. Either Redis or Memcached

serviceUpdate_serviceUpdateReleaseDate :: Lens' ServiceUpdate (Maybe UTCTime)

The date when the service update is initially available

serviceUpdate_autoUpdateAfterRecommendedApplyByDate :: Lens' ServiceUpdate (Maybe Bool)

Indicates whether the service update will be automatically applied once the recommended apply-by date has expired.

serviceUpdate_serviceUpdateSeverity :: Lens' ServiceUpdate (Maybe ServiceUpdateSeverity)

The severity of the service update

serviceUpdate_serviceUpdateEndDate :: Lens' ServiceUpdate (Maybe UTCTime)

The date after which the service update is no longer available

serviceUpdate_serviceUpdateDescription :: Lens' ServiceUpdate (Maybe Text)

Provides details of the service update

serviceUpdate_serviceUpdateRecommendedApplyByDate :: Lens' ServiceUpdate (Maybe UTCTime)

The recommendend date to apply the service update in order to ensure compliance. For information on compliance, see Self-Service Security Updates for Compliance.

serviceUpdate_serviceUpdateStatus :: Lens' ServiceUpdate (Maybe ServiceUpdateStatus)

The status of the service update

serviceUpdate_estimatedUpdateTime :: Lens' ServiceUpdate (Maybe Text)

The estimated length of time the service update will take