Network.AWS.ElastiCache.Types.Snapshot

Description

Documentation

data Snapshot

Represents a copy of an entire Redis cluster as of the time when the snapshot was taken.

See: newSnapshot smart constructor.

Constructors

Snapshot'

Fields

cacheClusterCreateTime :: Maybe ISO8601

The date and time when the source cluster was created.

nodeSnapshots :: Maybe [NodeSnapshot]

A list of the cache nodes in the source cluster.

automaticFailover :: Maybe AutomaticFailoverStatus

Indicates the status of automatic failover for the source Redis replication group.

numCacheNodes :: Maybe Int

The number of cache nodes in the source cluster.

For clusters running Redis, this value must be 1. For clusters running Memcached, this value must be between 1 and 40.

replicationGroupId :: Maybe Text

The unique identifier of the source replication group.

cacheClusterId :: Maybe Text

The user-supplied identifier of the source cluster.

snapshotStatus :: Maybe Text

The status of the snapshot. Valid values: creating | available | restoring | copying | deleting.

snapshotWindow :: Maybe Text

The daily time range during which ElastiCache takes daily snapshots of the source cluster.

cacheParameterGroupName :: Maybe Text

The cache parameter group that is associated with the source cluster.

replicationGroupDescription :: Maybe Text

A description of the source replication group.

snapshotRetentionLimit :: Maybe Int

For an automatic snapshot, the number of days for which ElastiCache retains the snapshot before deleting it.

For manual snapshots, this field reflects the SnapshotRetentionLimit for the source cluster when the snapshot was created. This field is otherwise ignored: Manual snapshots do not expire, and can only be deleted using the DeleteSnapshot operation.

Important If the value of SnapshotRetentionLimit is set to zero (0), backups are turned off.

arn :: Maybe Text

The ARN (Amazon Resource Name) of the snapshot.

preferredAvailabilityZone :: Maybe Text

The name of the Availability Zone in which the source cluster is located.

numNodeGroups :: Maybe Int

The number of node groups (shards) in this snapshot. When restoring from a snapshot, the number of node groups (shards) in the snapshot and in the restored replication group must be the same.

kmsKeyId :: Maybe Text

The ID of the KMS key used to encrypt the snapshot.

cacheSubnetGroupName :: Maybe Text

The name of the cache subnet group associated with the source cluster.

engineVersion :: Maybe Text

The version of the cache engine version that is used by the source cluster.

cacheNodeType :: Maybe Text

The name of the compute and memory capacity node type for the source cluster.

The following node types are supported by ElastiCache. Generally speaking, the current generation types provide more memory and computational power at lower cost when compared to their equivalent previous generation counterparts.

General purpose:

Current generation:

M6g node types (available only for Redis engine version 5.0.6 onward and for Memcached engine version 1.5.16 onward).

cache.m6g.large, cache.m6g.xlarge, cache.m6g.2xlarge, cache.m6g.4xlarge, cache.m6g.8xlarge, cache.m6g.12xlarge, cache.m6g.16xlarge

For region availability, see Supported Node Types

M5 node types: cache.m5.large, cache.m5.xlarge, cache.m5.2xlarge, cache.m5.4xlarge, cache.m5.12xlarge, cache.m5.24xlarge

M4 node types: cache.m4.large, cache.m4.xlarge, cache.m4.2xlarge, cache.m4.4xlarge, cache.m4.10xlarge

T3 node types: cache.t3.micro, cache.t3.small, cache.t3.medium

T2 node types: cache.t2.micro, cache.t2.small, cache.t2.medium

Previous generation: (not recommended)

T1 node types: cache.t1.micro

M1 node types: cache.m1.small, cache.m1.medium, cache.m1.large, cache.m1.xlarge

M3 node types: cache.m3.medium, cache.m3.large, cache.m3.xlarge, cache.m3.2xlarge

Compute optimized:

Previous generation: (not recommended)

C1 node types: cache.c1.xlarge

Memory optimized:

Current generation:

R6g node types (available only for Redis engine version 5.0.6 onward and for Memcached engine version 1.5.16 onward).

cache.r6g.large, cache.r6g.xlarge, cache.r6g.2xlarge, cache.r6g.4xlarge, cache.r6g.8xlarge, cache.r6g.12xlarge, cache.r6g.16xlarge

For region availability, see Supported Node Types

R5 node types: cache.r5.large, cache.r5.xlarge, cache.r5.2xlarge, cache.r5.4xlarge, cache.r5.12xlarge, cache.r5.24xlarge

R4 node types: cache.r4.large, cache.r4.xlarge, cache.r4.2xlarge, cache.r4.4xlarge, cache.r4.8xlarge, cache.r4.16xlarge

Previous generation: (not recommended)

M2 node types: cache.m2.xlarge, cache.m2.2xlarge, cache.m2.4xlarge

R3 node types: cache.r3.large, cache.r3.xlarge, cache.r3.2xlarge, cache.r3.4xlarge, cache.r3.8xlarge

Additional node type info

All current generation instance types are created in Amazon VPC by default.

Redis append-only files (AOF) are not supported for T1 or T2 instances.

Redis Multi-AZ with automatic failover is not supported on T1 instances.

Redis configuration variables appendonly and appendfsync are not supported on Redis version 2.8.22 and later.

preferredMaintenanceWindow :: Maybe Text

Specifies the weekly time range during which maintenance on the cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period.

Valid values for ddd are:

sun

mon

tue

wed

thu

fri

sat

Example: sun:23:00-mon:01:30

topicArn :: Maybe Text

The Amazon Resource Name (ARN) for the topic used by the source cluster for publishing notifications.

snapshotSource :: Maybe Text

Indicates whether the snapshot is from an automatic backup (automated) or was created manually (manual).

port :: Maybe Int

The port number used by each cache nodes in the source cluster.

preferredOutpostArn :: Maybe Text

The ARN (Amazon Resource Name) of the preferred outpost.

engine :: Maybe Text

The name of the cache engine (memcached or redis) used by the source cluster.

snapshotName :: Maybe Text

The name of a snapshot. For an automatic snapshot, the name is system-generated. For a manual snapshot, this is the user-provided name.

vpcId :: Maybe Text

The Amazon Virtual Private Cloud identifier (VPC ID) of the cache subnet group for the source cluster.

autoMinorVersionUpgrade :: Maybe Bool

This parameter is currently disabled.

newSnapshot :: Snapshot

Create a value of Snapshot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cacheClusterCreateTime:Snapshot', snapshot_cacheClusterCreateTime - The date and time when the source cluster was created.

$sel:nodeSnapshots:Snapshot', snapshot_nodeSnapshots - A list of the cache nodes in the source cluster.

$sel:automaticFailover:Snapshot', snapshot_automaticFailover - Indicates the status of automatic failover for the source Redis replication group.

$sel:numCacheNodes:Snapshot', snapshot_numCacheNodes - The number of cache nodes in the source cluster.

For clusters running Redis, this value must be 1. For clusters running Memcached, this value must be between 1 and 40.

$sel:replicationGroupId:Snapshot', snapshot_replicationGroupId - The unique identifier of the source replication group.

$sel:cacheClusterId:Snapshot', snapshot_cacheClusterId - The user-supplied identifier of the source cluster.

$sel:snapshotStatus:Snapshot', snapshot_snapshotStatus - The status of the snapshot. Valid values: creating | available | restoring | copying | deleting.

$sel:snapshotWindow:Snapshot', snapshot_snapshotWindow - The daily time range during which ElastiCache takes daily snapshots of the source cluster.

$sel:cacheParameterGroupName:Snapshot', snapshot_cacheParameterGroupName - The cache parameter group that is associated with the source cluster.

$sel:replicationGroupDescription:Snapshot', snapshot_replicationGroupDescription - A description of the source replication group.

$sel:snapshotRetentionLimit:Snapshot', snapshot_snapshotRetentionLimit - For an automatic snapshot, the number of days for which ElastiCache retains the snapshot before deleting it.

For manual snapshots, this field reflects the SnapshotRetentionLimit for the source cluster when the snapshot was created. This field is otherwise ignored: Manual snapshots do not expire, and can only be deleted using the DeleteSnapshot operation.

Important If the value of SnapshotRetentionLimit is set to zero (0), backups are turned off.

$sel:arn:Snapshot', snapshot_arn - The ARN (Amazon Resource Name) of the snapshot.

$sel:preferredAvailabilityZone:Snapshot', snapshot_preferredAvailabilityZone - The name of the Availability Zone in which the source cluster is located.

$sel:numNodeGroups:Snapshot', snapshot_numNodeGroups - The number of node groups (shards) in this snapshot. When restoring from a snapshot, the number of node groups (shards) in the snapshot and in the restored replication group must be the same.

$sel:kmsKeyId:Snapshot', snapshot_kmsKeyId - The ID of the KMS key used to encrypt the snapshot.

$sel:cacheSubnetGroupName:Snapshot', snapshot_cacheSubnetGroupName - The name of the cache subnet group associated with the source cluster.

$sel:engineVersion:Snapshot', snapshot_engineVersion - The version of the cache engine version that is used by the source cluster.

$sel:cacheNodeType:Snapshot', snapshot_cacheNodeType - The name of the compute and memory capacity node type for the source cluster.

The following node types are supported by ElastiCache. Generally speaking, the current generation types provide more memory and computational power at lower cost when compared to their equivalent previous generation counterparts.

General purpose:

Current generation:

M6g node types (available only for Redis engine version 5.0.6 onward and for Memcached engine version 1.5.16 onward).

cache.m6g.large, cache.m6g.xlarge, cache.m6g.2xlarge, cache.m6g.4xlarge, cache.m6g.8xlarge, cache.m6g.12xlarge, cache.m6g.16xlarge

For region availability, see Supported Node Types

M5 node types: cache.m5.large, cache.m5.xlarge, cache.m5.2xlarge, cache.m5.4xlarge, cache.m5.12xlarge, cache.m5.24xlarge

M4 node types: cache.m4.large, cache.m4.xlarge, cache.m4.2xlarge, cache.m4.4xlarge, cache.m4.10xlarge

T3 node types: cache.t3.micro, cache.t3.small, cache.t3.medium

T2 node types: cache.t2.micro, cache.t2.small, cache.t2.medium

Previous generation: (not recommended)

T1 node types: cache.t1.micro

M1 node types: cache.m1.small, cache.m1.medium, cache.m1.large, cache.m1.xlarge

M3 node types: cache.m3.medium, cache.m3.large, cache.m3.xlarge, cache.m3.2xlarge

Compute optimized:

Previous generation: (not recommended)

C1 node types: cache.c1.xlarge

Memory optimized:

Current generation:

R6g node types (available only for Redis engine version 5.0.6 onward and for Memcached engine version 1.5.16 onward).

cache.r6g.large, cache.r6g.xlarge, cache.r6g.2xlarge, cache.r6g.4xlarge, cache.r6g.8xlarge, cache.r6g.12xlarge, cache.r6g.16xlarge

For region availability, see Supported Node Types

R5 node types: cache.r5.large, cache.r5.xlarge, cache.r5.2xlarge, cache.r5.4xlarge, cache.r5.12xlarge, cache.r5.24xlarge

R4 node types: cache.r4.large, cache.r4.xlarge, cache.r4.2xlarge, cache.r4.4xlarge, cache.r4.8xlarge, cache.r4.16xlarge

Previous generation: (not recommended)

M2 node types: cache.m2.xlarge, cache.m2.2xlarge, cache.m2.4xlarge

R3 node types: cache.r3.large, cache.r3.xlarge, cache.r3.2xlarge, cache.r3.4xlarge, cache.r3.8xlarge

Additional node type info

All current generation instance types are created in Amazon VPC by default.

Redis append-only files (AOF) are not supported for T1 or T2 instances.

Redis Multi-AZ with automatic failover is not supported on T1 instances.

Redis configuration variables appendonly and appendfsync are not supported on Redis version 2.8.22 and later.

$sel:preferredMaintenanceWindow:Snapshot', snapshot_preferredMaintenanceWindow - Specifies the weekly time range during which maintenance on the cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period.

Valid values for ddd are:

sun

mon

tue

wed

thu

fri

sat

Example: sun:23:00-mon:01:30

$sel:topicArn:Snapshot', snapshot_topicArn - The Amazon Resource Name (ARN) for the topic used by the source cluster for publishing notifications.

$sel:snapshotSource:Snapshot', snapshot_snapshotSource - Indicates whether the snapshot is from an automatic backup (automated) or was created manually (manual).

$sel:port:Snapshot', snapshot_port - The port number used by each cache nodes in the source cluster.

$sel:preferredOutpostArn:Snapshot', snapshot_preferredOutpostArn - The ARN (Amazon Resource Name) of the preferred outpost.

$sel:engine:Snapshot', snapshot_engine - The name of the cache engine (memcached or redis) used by the source cluster.

$sel:snapshotName:Snapshot', snapshot_snapshotName - The name of a snapshot. For an automatic snapshot, the name is system-generated. For a manual snapshot, this is the user-provided name.

$sel:vpcId:Snapshot', snapshot_vpcId - The Amazon Virtual Private Cloud identifier (VPC ID) of the cache subnet group for the source cluster.

$sel:autoMinorVersionUpgrade:Snapshot', snapshot_autoMinorVersionUpgrade - This parameter is currently disabled.

snapshot_cacheClusterCreateTime :: Lens' Snapshot (Maybe UTCTime)

The date and time when the source cluster was created.

snapshot_nodeSnapshots :: Lens' Snapshot (Maybe [NodeSnapshot])

A list of the cache nodes in the source cluster.

snapshot_automaticFailover :: Lens' Snapshot (Maybe AutomaticFailoverStatus)

Indicates the status of automatic failover for the source Redis replication group.

snapshot_numCacheNodes :: Lens' Snapshot (Maybe Int)

The number of cache nodes in the source cluster.

For clusters running Redis, this value must be 1. For clusters running Memcached, this value must be between 1 and 40.

snapshot_replicationGroupId :: Lens' Snapshot (Maybe Text)

The unique identifier of the source replication group.

snapshot_cacheClusterId :: Lens' Snapshot (Maybe Text)

The user-supplied identifier of the source cluster.

snapshot_snapshotStatus :: Lens' Snapshot (Maybe Text)

The status of the snapshot. Valid values: creating | available | restoring | copying | deleting.

snapshot_snapshotWindow :: Lens' Snapshot (Maybe Text)

The daily time range during which ElastiCache takes daily snapshots of the source cluster.

snapshot_cacheParameterGroupName :: Lens' Snapshot (Maybe Text)

The cache parameter group that is associated with the source cluster.

snapshot_replicationGroupDescription :: Lens' Snapshot (Maybe Text)

A description of the source replication group.

snapshot_snapshotRetentionLimit :: Lens' Snapshot (Maybe Int)

For an automatic snapshot, the number of days for which ElastiCache retains the snapshot before deleting it.

For manual snapshots, this field reflects the SnapshotRetentionLimit for the source cluster when the snapshot was created. This field is otherwise ignored: Manual snapshots do not expire, and can only be deleted using the DeleteSnapshot operation.

Important If the value of SnapshotRetentionLimit is set to zero (0), backups are turned off.

snapshot_arn :: Lens' Snapshot (Maybe Text)

The ARN (Amazon Resource Name) of the snapshot.

snapshot_preferredAvailabilityZone :: Lens' Snapshot (Maybe Text)

The name of the Availability Zone in which the source cluster is located.

snapshot_numNodeGroups :: Lens' Snapshot (Maybe Int)

The number of node groups (shards) in this snapshot. When restoring from a snapshot, the number of node groups (shards) in the snapshot and in the restored replication group must be the same.

snapshot_kmsKeyId :: Lens' Snapshot (Maybe Text)

The ID of the KMS key used to encrypt the snapshot.

snapshot_cacheSubnetGroupName :: Lens' Snapshot (Maybe Text)

The name of the cache subnet group associated with the source cluster.

snapshot_engineVersion :: Lens' Snapshot (Maybe Text)

The version of the cache engine version that is used by the source cluster.

snapshot_cacheNodeType :: Lens' Snapshot (Maybe Text)

The name of the compute and memory capacity node type for the source cluster.

The following node types are supported by ElastiCache. Generally speaking, the current generation types provide more memory and computational power at lower cost when compared to their equivalent previous generation counterparts.

General purpose:

Current generation:

M6g node types (available only for Redis engine version 5.0.6 onward and for Memcached engine version 1.5.16 onward).

cache.m6g.large, cache.m6g.xlarge, cache.m6g.2xlarge, cache.m6g.4xlarge, cache.m6g.8xlarge, cache.m6g.12xlarge, cache.m6g.16xlarge

For region availability, see Supported Node Types

M5 node types: cache.m5.large, cache.m5.xlarge, cache.m5.2xlarge, cache.m5.4xlarge, cache.m5.12xlarge, cache.m5.24xlarge

M4 node types: cache.m4.large, cache.m4.xlarge, cache.m4.2xlarge, cache.m4.4xlarge, cache.m4.10xlarge

T3 node types: cache.t3.micro, cache.t3.small, cache.t3.medium

T2 node types: cache.t2.micro, cache.t2.small, cache.t2.medium

Previous generation: (not recommended)

T1 node types: cache.t1.micro

M1 node types: cache.m1.small, cache.m1.medium, cache.m1.large, cache.m1.xlarge

M3 node types: cache.m3.medium, cache.m3.large, cache.m3.xlarge, cache.m3.2xlarge

Compute optimized:

Previous generation: (not recommended)

C1 node types: cache.c1.xlarge

Memory optimized:

Current generation:

R6g node types (available only for Redis engine version 5.0.6 onward and for Memcached engine version 1.5.16 onward).

cache.r6g.large, cache.r6g.xlarge, cache.r6g.2xlarge, cache.r6g.4xlarge, cache.r6g.8xlarge, cache.r6g.12xlarge, cache.r6g.16xlarge

For region availability, see Supported Node Types

R5 node types: cache.r5.large, cache.r5.xlarge, cache.r5.2xlarge, cache.r5.4xlarge, cache.r5.12xlarge, cache.r5.24xlarge

R4 node types: cache.r4.large, cache.r4.xlarge, cache.r4.2xlarge, cache.r4.4xlarge, cache.r4.8xlarge, cache.r4.16xlarge

Previous generation: (not recommended)

M2 node types: cache.m2.xlarge, cache.m2.2xlarge, cache.m2.4xlarge

R3 node types: cache.r3.large, cache.r3.xlarge, cache.r3.2xlarge, cache.r3.4xlarge, cache.r3.8xlarge

Additional node type info

All current generation instance types are created in Amazon VPC by default.

Redis append-only files (AOF) are not supported for T1 or T2 instances.

Redis Multi-AZ with automatic failover is not supported on T1 instances.

Redis configuration variables appendonly and appendfsync are not supported on Redis version 2.8.22 and later.

snapshot_preferredMaintenanceWindow :: Lens' Snapshot (Maybe Text)

Specifies the weekly time range during which maintenance on the cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period.

Valid values for ddd are:

sun

mon

tue

wed

thu

fri

sat

Example: sun:23:00-mon:01:30

snapshot_topicArn :: Lens' Snapshot (Maybe Text)

The Amazon Resource Name (ARN) for the topic used by the source cluster for publishing notifications.

snapshot_snapshotSource :: Lens' Snapshot (Maybe Text)

Indicates whether the snapshot is from an automatic backup (automated) or was created manually (manual).

snapshot_port :: Lens' Snapshot (Maybe Int)

The port number used by each cache nodes in the source cluster.

snapshot_preferredOutpostArn :: Lens' Snapshot (Maybe Text)

The ARN (Amazon Resource Name) of the preferred outpost.

snapshot_engine :: Lens' Snapshot (Maybe Text)

The name of the cache engine (memcached or redis) used by the source cluster.

snapshot_snapshotName :: Lens' Snapshot (Maybe Text)

The name of a snapshot. For an automatic snapshot, the name is system-generated. For a manual snapshot, this is the user-provided name.

snapshot_vpcId :: Lens' Snapshot (Maybe Text)

The Amazon Virtual Private Cloud identifier (VPC ID) of the cache subnet group for the source cluster.

snapshot_autoMinorVersionUpgrade :: Lens' Snapshot (Maybe Bool)

This parameter is currently disabled.