Network.AWS.ElastiCache.Types.UpdateAction

Description

Documentation

data UpdateAction

The status of the service update for a specific replication group

See: newUpdateAction smart constructor.

Constructors

UpdateAction'

Fields

serviceUpdateSeverity :: Maybe ServiceUpdateSeverity

The severity of the service update

serviceUpdateReleaseDate :: Maybe ISO8601

The date the update is first available

replicationGroupId :: Maybe Text

The ID of the replication group

updateActionStatus :: Maybe UpdateActionStatus

The status of the update action

cacheClusterId :: Maybe Text

The ID of the cache cluster

serviceUpdateStatus :: Maybe ServiceUpdateStatus

The status of the service update

slaMet :: Maybe SlaMet

If yes, all nodes in the replication group have been updated by the recommended apply-by date. If no, at least one node in the replication group have not been updated by the recommended apply-by date. If N/A, the replication group was created after the recommended apply-by date.

serviceUpdateRecommendedApplyByDate :: Maybe ISO8601

The recommended date to apply the service update to ensure compliance. For information on compliance, see Self-Service Security Updates for Compliance.

updateActionAvailableDate :: Maybe ISO8601

The date that the service update is available to a replication group

nodeGroupUpdateStatus :: Maybe [NodeGroupUpdateStatus]

The status of the service update on the node group

serviceUpdateType :: Maybe ServiceUpdateType

Reflects the nature of the service update

cacheNodeUpdateStatus :: Maybe [CacheNodeUpdateStatus]

The status of the service update on the cache node

nodesUpdated :: Maybe Text

The progress of the service update on the replication group

engine :: Maybe Text

The Elasticache engine to which the update applies. Either Redis or Memcached

estimatedUpdateTime :: Maybe Text

The estimated length of time for the update to complete

updateActionStatusModifiedDate :: Maybe ISO8601

The date when the UpdateActionStatus was last modified

serviceUpdateName :: Maybe Text

The unique ID of the service update

newUpdateAction :: UpdateAction

Create a value of UpdateAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceUpdateSeverity:UpdateAction', updateAction_serviceUpdateSeverity - The severity of the service update

$sel:serviceUpdateReleaseDate:UpdateAction', updateAction_serviceUpdateReleaseDate - The date the update is first available

$sel:replicationGroupId:UpdateAction', updateAction_replicationGroupId - The ID of the replication group

$sel:updateActionStatus:UpdateAction', updateAction_updateActionStatus - The status of the update action

$sel:cacheClusterId:UpdateAction', updateAction_cacheClusterId - The ID of the cache cluster

$sel:serviceUpdateStatus:UpdateAction', updateAction_serviceUpdateStatus - The status of the service update

$sel:slaMet:UpdateAction', updateAction_slaMet - If yes, all nodes in the replication group have been updated by the recommended apply-by date. If no, at least one node in the replication group have not been updated by the recommended apply-by date. If N/A, the replication group was created after the recommended apply-by date.

$sel:serviceUpdateRecommendedApplyByDate:UpdateAction', updateAction_serviceUpdateRecommendedApplyByDate - The recommended date to apply the service update to ensure compliance. For information on compliance, see Self-Service Security Updates for Compliance.

$sel:updateActionAvailableDate:UpdateAction', updateAction_updateActionAvailableDate - The date that the service update is available to a replication group

$sel:nodeGroupUpdateStatus:UpdateAction', updateAction_nodeGroupUpdateStatus - The status of the service update on the node group

$sel:serviceUpdateType:UpdateAction', updateAction_serviceUpdateType - Reflects the nature of the service update

$sel:cacheNodeUpdateStatus:UpdateAction', updateAction_cacheNodeUpdateStatus - The status of the service update on the cache node

$sel:nodesUpdated:UpdateAction', updateAction_nodesUpdated - The progress of the service update on the replication group

$sel:engine:UpdateAction', updateAction_engine - The Elasticache engine to which the update applies. Either Redis or Memcached

$sel:estimatedUpdateTime:UpdateAction', updateAction_estimatedUpdateTime - The estimated length of time for the update to complete

$sel:updateActionStatusModifiedDate:UpdateAction', updateAction_updateActionStatusModifiedDate - The date when the UpdateActionStatus was last modified

$sel:serviceUpdateName:UpdateAction', updateAction_serviceUpdateName - The unique ID of the service update

updateAction_serviceUpdateSeverity :: Lens' UpdateAction (Maybe ServiceUpdateSeverity)

The severity of the service update

updateAction_serviceUpdateReleaseDate :: Lens' UpdateAction (Maybe UTCTime)

The date the update is first available

updateAction_replicationGroupId :: Lens' UpdateAction (Maybe Text)

The ID of the replication group

updateAction_updateActionStatus :: Lens' UpdateAction (Maybe UpdateActionStatus)

The status of the update action

updateAction_cacheClusterId :: Lens' UpdateAction (Maybe Text)

The ID of the cache cluster

updateAction_serviceUpdateStatus :: Lens' UpdateAction (Maybe ServiceUpdateStatus)

The status of the service update

updateAction_slaMet :: Lens' UpdateAction (Maybe SlaMet)

If yes, all nodes in the replication group have been updated by the recommended apply-by date. If no, at least one node in the replication group have not been updated by the recommended apply-by date. If N/A, the replication group was created after the recommended apply-by date.

updateAction_serviceUpdateRecommendedApplyByDate :: Lens' UpdateAction (Maybe UTCTime)

The recommended date to apply the service update to ensure compliance. For information on compliance, see Self-Service Security Updates for Compliance.

updateAction_updateActionAvailableDate :: Lens' UpdateAction (Maybe UTCTime)

The date that the service update is available to a replication group

updateAction_nodeGroupUpdateStatus :: Lens' UpdateAction (Maybe [NodeGroupUpdateStatus])

The status of the service update on the node group

updateAction_serviceUpdateType :: Lens' UpdateAction (Maybe ServiceUpdateType)

Reflects the nature of the service update

updateAction_cacheNodeUpdateStatus :: Lens' UpdateAction (Maybe [CacheNodeUpdateStatus])

The status of the service update on the cache node

updateAction_nodesUpdated :: Lens' UpdateAction (Maybe Text)

The progress of the service update on the replication group

updateAction_engine :: Lens' UpdateAction (Maybe Text)

The Elasticache engine to which the update applies. Either Redis or Memcached

updateAction_estimatedUpdateTime :: Lens' UpdateAction (Maybe Text)

The estimated length of time for the update to complete

updateAction_updateActionStatusModifiedDate :: Lens' UpdateAction (Maybe UTCTime)

The date when the UpdateActionStatus was last modified

updateAction_serviceUpdateName :: Lens' UpdateAction (Maybe Text)

The unique ID of the service update