Network.AWS.ElastiCache.V2014_03_24.AuthorizeCacheSecurityGroupIngress

Description

The AuthorizeCacheSecurityGroupIngress operation allows network ingress to a cache security group. Applications using ElastiCache must be running on Amazon EC2, and Amazon EC2 security groups are used as the authorization mechanism. You cannot authorize ingress from an Amazon EC2 security group in one Region to an ElastiCache cluster in another Region. https:/elasticache.us-east-1.amazonaws.com ?Action=AuthorizeCacheSecurityGroupIngress &EC2SecurityGroupName=default &CacheSecurityGroupName=mygroup &EC2SecurityGroupOwnerId=1234-5678-1234 &Version=2014-03-24 &SignatureVersion=2 &SignatureMethod=HmacSHA256 &Timestamp=2014-03-12T01%3A29%3A15.746Z &AWSAccessKeyId=YOUR-ACCESS-KEY &Signature=YOUR-SIGNATURE authorizing default 565419523791 mygroup 123456781234 My security group 817fa999-3647-11e0-ae57-f96cfe56749c.

Documentation

data AuthorizeCacheSecurityGroupIngress

Constructors

AuthorizeCacheSecurityGroupIngress

Fields

_acsgimCacheSecurityGroupName :: Text

The cache security group which will allow network ingress.

_acsgimEC2SecurityGroupOwnerId :: Text

The AWS account number of the Amazon EC2 security group owner. Note that this is not the same thing as an AWS access key ID - you must provide a valid AWS account number for this parameter.

_acsgimEC2SecurityGroupName :: Text

The Amazon EC2 security group to be authorized for ingress to the cache security group.

acsgimEC2SecurityGroupOwnerId :: Lens' AuthorizeCacheSecurityGroupIngress Text

acsgimEC2SecurityGroupName :: Lens' AuthorizeCacheSecurityGroupIngress Text

acsgimCacheSecurityGroupName :: Lens' AuthorizeCacheSecurityGroupIngress Text

data AuthorizeCacheSecurityGroupIngressResponse

Constructors

AuthorizeCacheSecurityGroupIngressResponse

Fields

_csgyCacheSecurityGroup :: Maybe CacheSecurityGroup

Represents the output of one of the following operations: AuthorizeCacheSecurityGroupIngress CreateCacheSecurityGroup RevokeCacheSecurityGroupIngress.

csgyCacheSecurityGroup :: Iso' AuthorizeCacheSecurityGroupIngressResponse (Maybe CacheSecurityGroup)