Network.AWS.ElastiCache.V2014_03_24.CopySnapshot

Description

The CopySnapshot operation makes a copy of an existing snapshot.

Documentation

data CopySnapshot

Constructors

CopySnapshot

Fields

_csmTargetSnapshotName :: Text

A name for the copied snapshot.

_csmSourceSnapshotName :: Text

The name of an existing snapshot from which to copy.

csmTargetSnapshotName :: Lens' CopySnapshot Text

csmSourceSnapshotName :: Lens' CopySnapshot Text

data CopySnapshotResponse

Constructors

CopySnapshotResponse

Fields

_sssssssssssssssrSnapshot :: Maybe Snapshot

Represents a copy of an entire cache cluster as of the time when the snapshot was taken.

sssssssssssssssrSnapshot :: Iso' CopySnapshotResponse (Maybe Snapshot)