Network.AWS.ElastiCache.V2014_03_24.CreateCacheParameterGroup

Description

The CreateCacheParameterGroup operation creates a new cache parameter group. A cache parameter group is a collection of parameters that you apply to all of the nodes in a cache cluster. https:/elasticache.us-east-1.amazonaws.com ?Action=CreateCacheParameterGroup &Description=My%20first%20cache%20parameter%20group &CacheParameterGroupFamily=memcached1.4 &CacheParameterGroupName=mycacheparametergroup1 &Version=2014-03-24 &SignatureVersion=2 &SignatureMethod=HmacSHA256 &Timestamp=2014-03-27T02%3A34%3A47.462Z &AWSAccessKeyId=YOUR-ACCESS-KEY &Signature=YOUR-SIGNATURE mycacheparametergroup3 memcached1.4 My first cache parameter group 05699541-b7f9-11e0-9326-b7275b9d4a6c.

Documentation

data CreateCacheParameterGroup

Constructors

CreateCacheParameterGroup

Fields

_ccpgmCacheParameterGroupFamily :: Text

The name of the cache parameter group family the cache parameter group can be used with. Valid values are: memcached1.4 | redis2.6 | redis2.8.

_ccpgmCacheParameterGroupName :: Text

A user-specified name for the cache parameter group.

_ccpgmDescription :: Text

A user-specified description for the cache parameter group.

ccpgmDescription :: Lens' CreateCacheParameterGroup Text

ccpgmCacheParameterGroupName :: Lens' CreateCacheParameterGroup Text

ccpgmCacheParameterGroupFamily :: Lens' CreateCacheParameterGroup Text

data CreateCacheParameterGroupResponse

Constructors

CreateCacheParameterGroupResponse

Fields

_cpgxCacheParameterGroup :: Maybe CacheParameterGroup

Represents the output of a CreateCacheParameterGroup operation.

cpgxCacheParameterGroup :: Iso' CreateCacheParameterGroupResponse (Maybe CacheParameterGroup)