Network.AWS.ElastiCache.V2014_03_24.CreateCacheSecurityGroup

Description

The CreateCacheSecurityGroup operation creates a new cache security group. Use a cache security group to control access to one or more cache clusters. Cache security groups are only used when you are creating a cluster outside of an Amazon Virtual Private Cloud (VPC). If you are creating a cluster inside of a VPC, use a cache subnet group instead. For more information, see CreateCacheSubnetGroup. https:/elasticache.us-east-1.amazonaws.com ?Action=CreateCacheSecurityGroup &CacheSecurityGroupName=mycachesecuritygroup &Description=My%20cache%20security%20group &Version=2014-03-24 &SignatureVersion=2 &SignatureMethod=HmacSHA256 &Timestamp=2014-03-27T02%3A43%3A10.703Z &AWSAccessKeyId=YOUR-ACCESS-KEY &Signature=YOUR-SIGNATURE mycachesecuritygroup 123456789012 My cache security group 2b1c8035-b7fa-11e0-9326-b7275b9d4a6c.

Documentation

data CreateCacheSecurityGroup

Constructors

CreateCacheSecurityGroup

Fields

_ccsgmCacheSecurityGroupName :: Text

A name for the cache security group. This value is stored as a lowercase string. Constraints: Must contain no more than 255 alphanumeric characters. Must not be the word Default. Example: mysecuritygroup.

_ccsgmDescription :: Text

A description for the cache security group.

ccsgmDescription :: Lens' CreateCacheSecurityGroup Text

ccsgmCacheSecurityGroupName :: Lens' CreateCacheSecurityGroup Text

data CreateCacheSecurityGroupResponse

Constructors

CreateCacheSecurityGroupResponse

Fields

_csgxCacheSecurityGroup :: Maybe CacheSecurityGroup

Represents the output of one of the following operations: AuthorizeCacheSecurityGroupIngress CreateCacheSecurityGroup RevokeCacheSecurityGroupIngress.

csgxCacheSecurityGroup :: Iso' CreateCacheSecurityGroupResponse (Maybe CacheSecurityGroup)