Network.AWS.ElastiCache.V2014_03_24.CreateSnapshot

Description

The CreateSnapshot operation creates a copy of an entire cache cluster at a specific moment in time.

Documentation

data CreateSnapshot

Constructors

CreateSnapshot

Fields

_csoCacheClusterId :: Text

The identifier of an existing cache cluster. The snapshot will be created from this cache cluster.

_csoSnapshotName :: Text

A name for the snapshot being created.

csoSnapshotName :: Lens' CreateSnapshot Text

csoCacheClusterId :: Lens' CreateSnapshot Text

data CreateSnapshotResponse

Constructors

CreateSnapshotResponse

Fields

_sssssssssssssssssssssrSnapshot :: Maybe Snapshot

Represents a copy of an entire cache cluster as of the time when the snapshot was taken.

sssssssssssssssssssssrSnapshot :: Iso' CreateSnapshotResponse (Maybe Snapshot)