Network.AWS.ElastiCache.V2014_03_24.DeleteCacheParameterGroup

Description

The DeleteCacheParameterGroup operation deletes the specified cache parameter group. You cannot delete a cache parameter group if it is associated with any cache clusters. https:/elasticache.us-east-1.amazonaws.com ?Action=DeleteCacheParameterGroup &CacheParameterGroupName=myparametergroup &Version=2014-03-24 &SignatureVersion=2 &SignatureMethod=HmacSHA256 &Timestamp=2014-03-25T21%3A16%3A39.166Z &AWSAccessKeyId=YOUR-ACCESS-KEY &Signature=YOUR-SIGNATURE d0a417cb-575b-11e0-8869-cd22b4f9d96f.

Documentation

data DeleteCacheParameterGroup

Constructors

DeleteCacheParameterGroup

Fields

_dcpgnCacheParameterGroupName :: Text

The name of the cache parameter group to delete. The specified cache security group must not be associated with any cache clusters.

dcpgnCacheParameterGroupName :: Iso' DeleteCacheParameterGroup Text

data DeleteCacheParameterGroupResponse

Constructors

DeleteCacheParameterGroupResponse