Network.AWS.ElastiCache.V2014_03_24.DeleteCacheSecurityGroup

Description

The DeleteCacheSecurityGroup operation deletes a cache security group. You cannot delete a cache security group if it is associated with any cache clusters. https:/elasticache.us-east-1.amazonaws.com ?Action=DeleteCacheSecurityGroup &CacheSecurityGroupName=mycachesecuritygroup3 &Version=2014-03-24 &SignatureVersion=2 &SignatureMethod=HmacSHA256 &Timestamp=2014-03-27T02%3A54%3A12.418Z &AWSAccessKeyId=YOUR-ACCESS-KEY &Signature=YOUR-SIGNATURE c130cfb7-3650-11e0-ae57-f96cfe56749c.

Documentation

data DeleteCacheSecurityGroup

Constructors

DeleteCacheSecurityGroup

Fields

_dcsgmCacheSecurityGroupName :: Text

The name of the cache security group to delete. You cannot delete the default security group.

dcsgmCacheSecurityGroupName :: Iso' DeleteCacheSecurityGroup Text

data DeleteCacheSecurityGroupResponse

Constructors

DeleteCacheSecurityGroupResponse