Network.AWS.ElastiCache.V2014_03_24.DeleteSnapshot

Description

The DeleteSnapshot operation deletes an existing snapshot. When you receive a successful response from this operation, ElastiCache immediately begins deleting the snapshot; you cannot cancel or revert this operation.

Documentation

data DeleteSnapshot

Constructors

DeleteSnapshot

Fields

_dsnSnapshotName :: Text

The name of the snapshot to be deleted.

dsnSnapshotName :: Iso' DeleteSnapshot Text

data DeleteSnapshotResponse

Constructors

DeleteSnapshotResponse

Fields

_sssssssssssssssssszSnapshot :: Maybe Snapshot

Represents a copy of an entire cache cluster as of the time when the snapshot was taken.

sssssssssssssssssszSnapshot :: Iso' DeleteSnapshotResponse (Maybe Snapshot)