Network.AWS.ElastiCache.V2014_03_24.DescribeCacheParameterGroups

Description

The DescribeCacheParameterGroups operation returns a list of cache parameter group descriptions. If a cache parameter group name is specified, the list will contain only the descriptions for that group. https:/elasticache.us-east-1.amazonaws.com ?Action=DescribeCacheParameterGroups &MaxRecords=100 &Version=2014-03-24 &SignatureVersion=2 &SignatureMethod=HmacSHA256 &Timestamp=2014-03-27T01%3A34%3A31.045Z &AWSAccessKeyId=YOUR-ACCESS-KEY &Signature=YOUR-SIGNATURE default.memcached1.4 memcached1.4 Default parameter group for memcached1.4 mycacheparametergroup memcached1.4 My cache parameter group mycacheparametergroup1 memcached1.4 My first cache parameter group mycacheparametergroup3 memcached1.4 My first cache parameter group 7193fbb8-b7fc-11e0-9b0b-a9261be2b354.

Documentation

describeCacheParameterGroups :: DescribeCacheParameterGroups

Minimum specification for a DescribeCacheParameterGroups request.

data DescribeCacheParameterGroups

Constructors

DescribeCacheParameterGroups

Fields

_dcpgmMaxRecords :: Maybe Integer

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: minimum 20; maximum 100.

_dcpgmCacheParameterGroupName :: Maybe Text

The name of a specific cache parameter group to return details for.

_dcpgmMarker :: Maybe Text

An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

dcpgmMaxRecords :: Lens' DescribeCacheParameterGroups (Maybe Integer)

dcpgmMarker :: Lens' DescribeCacheParameterGroups (Maybe Text)

dcpgmCacheParameterGroupName :: Lens' DescribeCacheParameterGroups (Maybe Text)

data DescribeCacheParameterGroupsResponse

Constructors

DescribeCacheParameterGroupsResponse

Fields

_cpgmCacheParameterGroups :: [CacheParameterGroup]

A list of cache parameter groups. Each element in the list contains detailed information about one cache parameter group.

_cpgmMarker :: Maybe Text

Provides an identifier to allow retrieval of paginated results.

cpgmMarker :: Lens' DescribeCacheParameterGroupsResponse (Maybe Text)

cpgmCacheParameterGroups :: Lens' DescribeCacheParameterGroupsResponse [CacheParameterGroup]