Network.AWS.ElastiCache.V2014_03_24.ModifyCacheParameterGroup

Description

The ModifyCacheParameterGroup operation modifies the parameters of a cache parameter group. You can modify up to 20 parameters in a single request by submitting a list parameter name and value pairs. https:/elasticache.us-east-1.amazonaws.com ?Action=ModifyCacheParameterGroup ?ParameterNameValues.member.1.ParameterName=chunk_size_growth_factor &ParameterNameValues.member.1.ParameterValue=1.02 &CacheParameterGroupName=mycacheparametergroup &Version=2014-03-24 &SignatureVersion=2 &SignatureMethod=HmacSHA256 &Timestamp=2014-03-27T03%3A24%3A50.203Z &AWSAccessKeyId=YOUR-ACCESS-KEY &Signature=YOUR-SIGNATURE mycacheparametergroup fcedeef2-b7ff-11e0-9326-b7275b9d4a6c.

Documentation

data ModifyCacheParameterGroup

Constructors

ModifyCacheParameterGroup

Fields

_mcpgmParameterNameValues :: [ParameterNameValue]

An array of parameter names and values for the parameter update. You must supply at least one parameter name and value; subsequent arguments are optional. A maximum of 20 parameters may be modified per request.

_mcpgmCacheParameterGroupName :: Text

The name of the cache parameter group to modify.

mcpgmParameterNameValues :: Lens' ModifyCacheParameterGroup [ParameterNameValue]

mcpgmCacheParameterGroupName :: Lens' ModifyCacheParameterGroup Text

data ModifyCacheParameterGroupResponse

Constructors

ModifyCacheParameterGroupResponse

Fields

_cpgnmCacheParameterGroupName :: Maybe Text

The name of the cache parameter group.

cpgnmCacheParameterGroupName :: Iso' ModifyCacheParameterGroupResponse (Maybe Text)