Network.AWS.ElastiCache.V2014_03_24.PurchaseReservedCacheNodesOffering

Description

The PurchaseReservedCacheNodesOffering operation allows you to purchase a reserved cache node offering. https:/elasticache.amazonaws.com ?Action=PurchaseReservedCacheNodesOffering &ReservedCacheNodeId=myreservationID &ReservedCacheNodesOfferingId=438012d3-4052-4cc7-b2e3-8d3372e0e706 &CacheNodeCount=1 &SignatureVersion=2 &SignatureMethod=HmacSHA256 &Timestamp=2014-03-10T18%3A31%3A36.118Z &AWSAccessKeyId= &Signature= Medium Utilization memcached 438012d3-4052-4cc7-b2e3-8d3372e0e706 payment-pending myreservationID 10 2014-03-18T23:24:56.577Z 31536000 123.0 0.123 cache.m1.small 7f099901-29cf-11e1-bd06-6fe008f046c3.

Documentation

purchaseReservedCacheNodesOffering

Arguments

:: Text
_prcnomReservedCacheNodesOfferingId

-> PurchaseReservedCacheNodesOffering

Minimum specification for a PurchaseReservedCacheNodesOffering request.

data PurchaseReservedCacheNodesOffering

Constructors

PurchaseReservedCacheNodesOffering

Fields

_prcnomReservedCacheNodesOfferingId :: Text

The ID of the reserved cache node offering to purchase. Example: 438012d3-4052-4cc7-b2e3-8d3372e0e706.

_prcnomCacheNodeCount :: Maybe Integer

The number of cache node instances to reserve. Default: 1.

_prcnomReservedCacheNodeId :: Maybe Text

A customer-specified identifier to track this reservation. Example: myreservationID.

prcnomReservedCacheNodesOfferingId :: Lens' PurchaseReservedCacheNodesOffering Text

prcnomReservedCacheNodeId :: Lens' PurchaseReservedCacheNodesOffering (Maybe Text)

prcnomCacheNodeCount :: Lens' PurchaseReservedCacheNodesOffering (Maybe Integer)

data PurchaseReservedCacheNodesOfferingResponse

Constructors

PurchaseReservedCacheNodesOfferingResponse

Fields

_rcnwReservedCacheNode :: Maybe ReservedCacheNode

Represents the output of a PurchaseReservedCacheNodesOffering operation.

rcnwReservedCacheNode :: Iso' PurchaseReservedCacheNodesOfferingResponse (Maybe ReservedCacheNode)