Network.AWS.ElastiCache.V2014_03_24.RebootCacheCluster

Description

The RebootCacheCluster operation reboots some, or all, of the cache cluster nodes within a provisioned cache cluster. This API will apply any modified cache parameter groups to the cache cluster. The reboot action takes place as soon as possible, and results in a momentary outage to the cache cluster. During the reboot, the cache cluster status is set to REBOOTING. The reboot causes the contents of the cache (for each cache cluster node being rebooted) to be lost. When the reboot is complete, a cache cluster event is created.

Documentation

data RebootCacheCluster

Constructors

RebootCacheCluster

Fields

_rccmCacheNodeIdsToReboot :: [Text]

A list of cache cluster node IDs to reboot. A node ID is a numeric identifier (0001, 0002, etc.). To reboot an entire cache cluster, specify all of the cache cluster node IDs.

_rccmCacheClusterId :: Text

The cache cluster identifier. This parameter is stored as a lowercase string.

rccmCacheNodeIdsToReboot :: Lens' RebootCacheCluster [Text]

rccmCacheClusterId :: Lens' RebootCacheCluster Text

data RebootCacheClusterResponse

Constructors

RebootCacheClusterResponse

Fields

_ccxCacheCluster :: Maybe CacheCluster

Contains all of the attributes of a specific cache cluster.

ccxCacheCluster :: Iso' RebootCacheClusterResponse (Maybe CacheCluster)