Network.AWS.ElastiCache.V2014_03_24.ResetCacheParameterGroup

Description

The ResetCacheParameterGroup operation modifies the parameters of a cache parameter group to the engine or system default value. You can reset specific parameters by submitting a list of parameter names. To reset the entire cache parameter group, specify the ResetAllParameters and CacheParameterGroupName parameters.

Documentation

resetCacheParameterGroup

Arguments

:: [ParameterNameValue]
_rcpgmParameterNameValues

-> Text
_rcpgmCacheParameterGroupName

-> ResetCacheParameterGroup

Minimum specification for a ResetCacheParameterGroup request.

data ResetCacheParameterGroup

Constructors

ResetCacheParameterGroup

Fields

_rcpgmParameterNameValues :: [ParameterNameValue]

An array of parameter names to be reset. If you are not resetting the entire cache parameter group, you must specify at least one parameter name.

_rcpgmCacheParameterGroupName :: Text

The name of the cache parameter group to reset.

_rcpgmResetAllParameters :: Maybe Bool

If true, all parameters in the cache parameter group will be reset to default values. If false, no such action occurs. Valid values: true | false.

rcpgmResetAllParameters :: Lens' ResetCacheParameterGroup (Maybe Bool)

rcpgmParameterNameValues :: Lens' ResetCacheParameterGroup [ParameterNameValue]

rcpgmCacheParameterGroupName :: Lens' ResetCacheParameterGroup Text

data ResetCacheParameterGroupResponse

Constructors

ResetCacheParameterGroupResponse

Fields

_cpgnnCacheParameterGroupName :: Maybe Text

The name of the cache parameter group.

cpgnnCacheParameterGroupName :: Iso' ResetCacheParameterGroupResponse (Maybe Text)