Network.AWS.ElastiCache.V2014_03_24.RevokeCacheSecurityGroupIngress

Description

The RevokeCacheSecurityGroupIngress operation revokes ingress from a cache security group. Use this operation to disallow access from an Amazon EC2 security group that had been previously authorized. https:/elasticache.us-east-1.amazonaws.com ?Action=RevokeCacheSecurityGroupIngress &EC2SecurityGroupName=default &CacheSecurityGroupName=mygroup &EC2SecurityGroupOwnerId=1234-5678-1234 &Version=2014-03-24 &SignatureVersion=2 &SignatureMethod=HmacSHA256 &Timestamp=2014-03-27T02%3A30%3A08.444Z &AWSAccessKeyId=YOUR-ACCESS-KEY &Signature=YOUR-SIGNATURE revoking default 123456781234 mygroup 123456789012 My security group 02ae3699-3650-11e0-a564-8f11342c56b0.

Documentation

data RevokeCacheSecurityGroupIngress

Constructors

RevokeCacheSecurityGroupIngress

Fields

_rcsgimCacheSecurityGroupName :: Text

The name of the cache security group to revoke ingress from.

_rcsgimEC2SecurityGroupOwnerId :: Text

The AWS account number of the Amazon EC2 security group owner. Note that this is not the same thing as an AWS access key ID - you must provide a valid AWS account number for this parameter.

_rcsgimEC2SecurityGroupName :: Text

The name of the Amazon EC2 security group to revoke access from.

rcsgimEC2SecurityGroupOwnerId :: Lens' RevokeCacheSecurityGroupIngress Text

rcsgimEC2SecurityGroupName :: Lens' RevokeCacheSecurityGroupIngress Text

rcsgimCacheSecurityGroupName :: Lens' RevokeCacheSecurityGroupIngress Text

data RevokeCacheSecurityGroupIngressResponse

Constructors

RevokeCacheSecurityGroupIngressResponse

Fields

_csgwCacheSecurityGroup :: Maybe CacheSecurityGroup

Represents the output of one of the following operations: AuthorizeCacheSecurityGroupIngress CreateCacheSecurityGroup RevokeCacheSecurityGroupIngress.

csgwCacheSecurityGroup :: Iso' RevokeCacheSecurityGroupIngressResponse (Maybe CacheSecurityGroup)