Network.AWS.ElastiCache.V2014_03_24.Types

Description

ElastiCache is a web service that makes it easy to deploy, operate, and scale an in-memory cache in the cloud. The service improves the performance of web applications by allowing you to retrieve information from fast, managed, in-memory caches, instead of relying entirely on slower disk-based databases. Amazon ElastiCache automatically detects and replaces failed nodes, reducing the overhead associated with self-managed infrastructures and provides a resilient system that mitigates the risk of overloaded databases, which slow website and application load times. Through integration with Amazon CloudWatch, Amazon ElastiCache provides enhanced visibility into key performance metrics associated with your Memcached or Redis nodes.

Documentation

data ElastiCache

Supported version (2014-03-24) of the Amazon ElastiCache service.

xmlOptions :: Tagged a XMLOptions

data SourceType

The event source to retrieve events for. If no value is specified, all events are returned. Valid values are: cache-cluster | cache-parameter-group | cache-security-group | cache-subnet-group.

Constructors

SourceTypeCacheCluster
cache-cluster

SourceTypeCacheParameterGroup
cache-parameter-group

SourceTypeCacheSecurityGroup
cache-security-group

SourceTypeCacheSubnetGroup
cache-subnet-group

newtype AvailabilityZone

The Availability Zone associated with the subnet.

Constructors

AvailabilityZone

Fields

_azName :: Maybe Text

The name of the availability zone.

newtype ReplicationGroupPendingModifiedValues

A group of settings to be applied to the replication group, either immediately or during the next maintenance window.

Constructors

ReplicationGroupPendingModifiedValues

Fields

_rgpmvPrimaryClusterId :: Maybe Text

The primary cluster ID which will be applied immediately (if --apply-immediately was specified), or during the next maintenance window.

data CacheCluster

Contains all of the attributes of a specific cache cluster.

Constructors

CacheCluster

Fields

_ccEngineVersion :: Maybe Text

The version of the cache engine version that is used in this cache cluster.

_ccCacheNodeType :: Maybe Text

The name of the compute and memory capacity node type for the cache cluster.

_ccCacheNodes :: [CacheNode]

A list of cache nodes that are members of the cache cluster.

_ccCacheClusterCreateTime :: Maybe ISO8601

The date and time when the cache cluster was created.

_ccAutoMinorVersionUpgrade :: Maybe Bool

If true, then minor version patches are applied automatically; if false, then automatic minor version patches are disabled.

_ccSecurityGroups :: [SecurityGroupMembership]

A list of VPC Security Groups associated with the cache cluster.

_ccNotificationConfiguration :: Maybe NotificationConfiguration

Describes a notification topic and its status. Notification topics are used for publishing ElastiCache events to subscribers using Amazon Simple Notification Service (SNS).

_ccSnapshotWindow :: Maybe Text

The daily time range (in UTC) during which ElastiCache will begin taking a daily snapshot of your cache cluster. Example: 05:00-09:00.

_ccCacheClusterId :: Maybe Text

The user-supplied identifier of the cache cluster. This is a unique key that identifies a cache cluster.

_ccConfigurationEndpoint :: Maybe Endpoint

Represents the information required for client programs to connect to a cache node.

_ccEngine :: Maybe Text

The name of the cache engine (memcached or redis) to be used for this cache cluster.

_ccCacheSecurityGroups :: [CacheSecurityGroupMembership]

A list of cache security group elements, composed of name and status sub-elements.

_ccClientDownloadLandingPage :: Maybe Text

The URL of the web page where you can download the latest ElastiCache client library.

_ccPreferredMaintenanceWindow :: Maybe Text

The time range (in UTC) during which weekly system maintenance can occur.

_ccCacheSubnetGroupName :: Maybe Text

The name of the cache subnet group associated with the cache cluster.

_ccPreferredAvailabilityZone :: Maybe Text

The name of the Availability Zone in which the cache cluster is located.

_ccCacheParameterGroup :: Maybe CacheParameterGroupStatus

The status of the cache parameter group.

_ccCacheClusterStatus :: Maybe Text

The current state of this cache cluster - creating, available, etc.

_ccSnapshotRetentionLimit :: Maybe Integer

The number of days for which ElastiCache will retain automatic cache cluster snapshots before deleting them. For example, if you set SnapshotRetentionLimit to 5, then a snapshot that was taken today will be retained for 5 days before being deleted.

_ccReplicationGroupId :: Maybe Text

The replication group to which this cache cluster belongs. If this field is empty, the cache cluster is not associated with any replication group.

_ccPendingModifiedValues :: Maybe PendingModifiedValues

A group of settings that will be applied to the cache cluster in the future, or that are currently being applied.

_ccNumCacheNodes :: Maybe Integer

The number of cache nodes in the cache cluster.

data CacheEngineVersion

Provides all of the details about a particular cache engine version.

Constructors

CacheEngineVersion

Fields

_cevEngineVersion :: Maybe Text

The version number of the cache engine.

_cevCacheParameterGroupFamily :: Maybe Text

The name of the cache parameter group family associated with this cache engine.

_cevCacheEngineDescription :: Maybe Text

The description of the cache engine.

_cevEngine :: Maybe Text

The name of the cache engine.

_cevCacheEngineVersionDescription :: Maybe Text

The description of the cache engine version.

data CacheNode

Represents an individual cache node within a cache cluster. Each cache node runs its own instance of the cluster's protocol-compliant caching software - either Memcached or Redis.

Constructors

CacheNode

Fields

_cnSourceCacheNodeId :: Maybe Text

The ID of the primary node to which this read replica node is synchronized. If this field is empty, then this node is not associated with a primary cache cluster.

_cnParameterGroupStatus :: Maybe Text

The status of the parameter group applied to this cache node.

_cnCacheNodeCreateTime :: Maybe ISO8601

The date and time when the cache node was created.

_cnCacheNodeId :: Maybe Text

The cache node identifier. A node ID is a numeric identifier (0001, 0002, etc.). The combination of cluster ID and node ID uniquely identifies every cache node used in a customer's AWS account.

_cnCacheNodeStatus :: Maybe Text

The current state of this cache node.

_cnEndpoint :: Maybe Endpoint

The hostname and IP address for connecting to this cache node.

data CacheNodeTypeSpecificParameter

A parameter that has a different value for each cache node type it is applied to. For example, in a Redis cache cluster, a cache.m1.large cache node type would have a larger maxmemory value than a cache.m1.small type.

Constructors

CacheNodeTypeSpecificParameter

Fields

_cntspCacheNodeTypeSpecificValues :: [CacheNodeTypeSpecificValue]

A list of cache node types and their corresponding values for this parameter.

_cntspMinimumEngineVersion :: Maybe Text

The earliest cache engine version to which the parameter can apply.

_cntspSource :: Maybe Text

The source of the parameter value.

_cntspIsModifiable :: Maybe Bool

Indicates whether (true) or not (false) the parameter can be modified. Some parameters have security or operational implications that prevent them from being changed.

_cntspDataType :: Maybe Text

The valid data type for the parameter.

_cntspAllowedValues :: Maybe Text

The valid range of values for the parameter.

_cntspParameterName :: Maybe Text

The name of the parameter.

_cntspDescription :: Maybe Text

A description of the parameter.

data CacheNodeTypeSpecificValue

A value that applies only to a certain cache node type.

Constructors

CacheNodeTypeSpecificValue

Fields

_cntsvCacheNodeType :: Maybe Text

The cache node type for which this value applies.

_cntsvValue :: Maybe Text

The value for the cache node type.

data CacheParameterGroup

Represents the output of a CreateCacheParameterGroup operation.

Constructors

CacheParameterGroup

Fields

_cpgCacheParameterGroupFamily :: Maybe Text

The name of the cache parameter group family that this cache parameter group is compatible with.

_cpgCacheParameterGroupName :: Maybe Text

The name of the cache parameter group.

_cpgDescription :: Maybe Text

The description for this cache parameter group.

data CacheParameterGroupStatus

The status of the cache parameter group.

Constructors

CacheParameterGroupStatus

Fields

_cpgsCacheParameterGroupName :: Maybe Text

The name of the cache parameter group.

_cpgsCacheNodeIdsToReboot :: [Text]

A list of the cache node IDs which need to be rebooted for parameter changes to be applied. A node ID is a numeric identifier (0001, 0002, etc.).

_cpgsParameterApplyStatus :: Maybe Text

The status of parameter updates.

data CacheSecurityGroup

Represents the output of one of the following operations: AuthorizeCacheSecurityGroupIngress CreateCacheSecurityGroup RevokeCacheSecurityGroupIngress.

Constructors

CacheSecurityGroup

Fields

_csgCacheSecurityGroupName :: Maybe Text

The name of the cache security group.

_csgOwnerId :: Maybe Text

The AWS account ID of the cache security group owner.

_csgEC2SecurityGroups :: [EC2SecurityGroup]

A list of Amazon EC2 security groups that are associated with this cache security group.

_csgDescription :: Maybe Text

The description of the cache security group.

data CacheSecurityGroupMembership

Represents a cache cluster's status within a particular cache security group.

Constructors

CacheSecurityGroupMembership

Fields

_csgmStatus :: Maybe Text

The membership status in the cache security group. The status changes when a cache security group is modified, or when the cache security groups assigned to a cache cluster are modified.

_csgmCacheSecurityGroupName :: Maybe Text

The name of the cache security group.

data CacheSubnetGroup

Represents the output of one of the following operations: CreateCacheSubnetGroup ModifyCacheSubnetGroup.

Constructors

CacheSubnetGroup

Fields

_csiVpcId :: Maybe Text

The Amazon Virtual Private Cloud identifier (VPC ID) of the cache subnet group.

_csiSubnets :: [Subnet]

A list of subnets associated with the cache subnet group.

_csiCacheSubnetGroupName :: Maybe Text

The name of the cache subnet group.

_csiCacheSubnetGroupDescription :: Maybe Text

The description of the cache subnet group.

data EC2SecurityGroup

Provides ownership and status information for an Amazon EC2 security group.

Constructors

EC2SecurityGroup

Fields

_ecsgStatus :: Maybe Text

The status of the Amazon EC2 security group.

_ecsgEC2SecurityGroupOwnerId :: Maybe Text

The AWS account ID of the Amazon EC2 security group owner.

_ecsgEC2SecurityGroupName :: Maybe Text

The name of the Amazon EC2 security group.

data Endpoint

Represents the information required for client programs to connect to a cache node.

Constructors

Endpoint

Fields

_eAddress :: Maybe Text

The DNS hostname of the cache node.

_ePort :: Maybe Integer

The port number that the cache engine is listening on.

data EngineDefaults

Represents the output of a DescribeEngineDefaultParameters operation.

Constructors

EngineDefaults

Fields

_edCacheParameterGroupFamily :: Maybe Text

Specifies the name of the cache parameter group family to which the engine default parameters apply.

_edCacheNodeTypeSpecificParameters :: [CacheNodeTypeSpecificParameter]

A list of parameters specific to a particular cache node type. Each element in the list contains detailed information about one parameter.

_edMarker :: Maybe Text

Provides an identifier to allow retrieval of paginated results.

_edParameters :: [Parameter]

Contains a list of engine default parameters.

data Event

Represents a single occurrence of something interesting within the system. Some examples of events are creating a cache cluster, adding or removing a cache node, or rebooting a node.

Constructors

Event

Fields

_euSourceType :: Maybe SourceType

Specifies the origin of this event - a cache cluster, a parameter group, a security group, etc.

_euSourceIdentifier :: Maybe Text

The identifier for the source of the event. For example, if the event occurred at the cache cluster level, the identifier would be the name of the cache cluster.

_euDate :: Maybe ISO8601

The date and time when the event occurred.

_euMessage :: Maybe Text

The text of the event.

data NodeGroup

Represents a collection of cache nodes in a replication group.

Constructors

NodeGroup

Fields

_ngStatus :: Maybe Text

The current state of this replication group - creating, available, etc.

_ngPrimaryEndpoint :: Maybe Endpoint

Represents the information required for client programs to connect to a cache node.

_ngNodeGroupMembers :: [NodeGroupMember]

A list containing information about individual nodes within the node group.

_ngNodeGroupId :: Maybe Text

The identifier for the node group. A replication group contains only one node group; therefore, the node group ID is 0001.

data NodeGroupMember

Represents a single node within a node group.

Constructors

NodeGroupMember

Fields

_ngmCacheClusterId :: Maybe Text

The ID of the cache cluster to which the node belongs.

_ngmCacheNodeId :: Maybe Text

The ID of the node within its cache cluster. A node ID is a numeric identifier (0001, 0002, etc.).

_ngmPreferredAvailabilityZone :: Maybe Text

The name of the Availability Zone in which the node is located.

_ngmCurrentRole :: Maybe Text

The role that is currently assigned to the node - primary or replica.

_ngmReadEndpoint :: Maybe Endpoint

Represents the information required for client programs to connect to a cache node.

data NodeSnapshot

Represents an individual cache node in a snapshot of a cache cluster.

Constructors

NodeSnapshot

Fields

_nsCacheNodeCreateTime :: Maybe ISO8601

The date and time when the cache node was created in the source cache cluster.

_nsCacheNodeId :: Maybe Text

The cache node identifier for the node in the source cache cluster.

_nsSnapshotCreateTime :: Maybe ISO8601

The date and time when the source node's metadata and cache data set was obtained for the snapshot.

_nsCacheSize :: Maybe Text

The size of the cache on the source cache node.

data NotificationConfiguration

Describes a notification topic and its status. Notification topics are used for publishing ElastiCache events to subscribers using Amazon Simple Notification Service (SNS).

Constructors

NotificationConfiguration

Fields

_ncTopicStatus :: Maybe Text

The current state of the topic.

_ncTopicArn :: Maybe Text

The Amazon Resource Name (ARN) that identifies the topic.

data Parameter

Describes an individual setting that controls some aspect of ElastiCache behavior.

Constructors

Parameter

Fields

_prParameterValue :: Maybe Text

The value of the parameter.

_prMinimumEngineVersion :: Maybe Text

The earliest cache engine version to which the parameter can apply.

_prSource :: Maybe Text

The source of the parameter.

_prIsModifiable :: Maybe Bool

Indicates whether (true) or not (false) the parameter can be modified. Some parameters have security or operational implications that prevent them from being changed.

_prDataType :: Maybe Text

The valid data type for the parameter.

_prAllowedValues :: Maybe Text

The valid range of values for the parameter.

_prParameterName :: Maybe Text

The name of the parameter.

_prDescription :: Maybe Text

A description of the parameter.

data ParameterNameValue

Describes a name-value pair that is used to update the value of a parameter.

Constructors

ParameterNameValue

Fields

_pnvParameterValue :: Maybe Text

The value of the parameter.

_pnvParameterName :: Maybe Text

The name of the parameter.

data PendingModifiedValues

A group of settings that will be applied to the cache cluster in the future, or that are currently being applied.

Constructors

PendingModifiedValues

Fields

_pmvEngineVersion :: Maybe Text

The new cache engine version that the cache cluster will run.

_pmvCacheNodeIdsToRemove :: [Text]

A list of cache node IDs that are being removed (or will be removed) from the cache cluster. A node ID is a numeric identifier (0001, 0002, etc.).

_pmvNumCacheNodes :: Maybe Integer

The new number of cache nodes for the cache cluster.

data RecurringCharge

Contains the specific price and frequency of a recurring charges for a reserved cache node, or for a reserved cache node offering.

Constructors

RecurringCharge

Fields

_rcRecurringChargeFrequency :: Maybe Text

The frequency of the recurring charge.

_rcRecurringChargeAmount :: Maybe Double

The monetary amount of the recurring charge.

data ReplicationGroup

Contains all of the attributes of a specific replication group.

Constructors

ReplicationGroup

Fields

_rgStatus :: Maybe Text

The current state of this replication group - creating, available, etc.

_rgNodeGroups :: [NodeGroup]

A single element list with information about the nodes in the replication group.

_rgSnapshottingClusterId :: Maybe Text

The cache cluster ID that is used as the daily snapshot source for the replication group.

_rgMemberClusters :: [Text]

The names of all the cache clusters that are part of this replication group.

_rgDescription :: Maybe Text

The description of the replication group.

_rgReplicationGroupId :: Maybe Text

The identifier for the replication group.

_rgPendingModifiedValues :: Maybe ReplicationGroupPendingModifiedValues

A group of settings to be applied to the replication group, either immediately or during the next maintenance window.

data ReservedCacheNode

Represents the output of a PurchaseReservedCacheNodesOffering operation.

Constructors

ReservedCacheNode

Fields

_rcnCacheNodeType :: Maybe Text

The cache node type for the reserved cache nodes.

_rcnState :: Maybe Text

The state of the reserved cache node.

_rcnStartTime :: Maybe ISO8601

The time the reservation started.

_rcnProductDescription :: Maybe Text

The description of the reserved cache node.

_rcnCacheNodeCount :: Maybe Integer

The number of cache nodes that have been reserved.

_rcnReservedCacheNodeId :: Maybe Text

The unique identifier for the reservation.

_rcnRecurringCharges :: [RecurringCharge]

The recurring price charged to run this reserved cache node.

_rcnOfferingType :: Maybe Text

The offering type of this reserved cache node.

_rcnUsagePrice :: Maybe Double

The hourly price charged for this reserved cache node.

_rcnFixedPrice :: Maybe Double

The fixed price charged for this reserved cache node.

_rcnDuration :: Maybe Integer

The duration of the reservation in seconds.

_rcnReservedCacheNodesOfferingId :: Maybe Text

The offering identifier.

data ReservedCacheNodesOffering

Describes all of the attributes of a reserved cache node offering.

Constructors

ReservedCacheNodesOffering

Fields

_rcnoCacheNodeType :: Maybe Text

The cache node type for the reserved cache node.

_rcnoProductDescription :: Maybe Text

The cache engine used by the offering.

_rcnoRecurringCharges :: [RecurringCharge]

The recurring price charged to run this reserved cache node.

_rcnoOfferingType :: Maybe Text

The offering type.

_rcnoUsagePrice :: Maybe Double

The hourly price charged for this offering.

_rcnoFixedPrice :: Maybe Double

The fixed price charged for this offering.

_rcnoDuration :: Maybe Integer

The duration of the offering. in seconds.

_rcnoReservedCacheNodesOfferingId :: Maybe Text

A unique identifier for the reserved cache node offering.

data SecurityGroupMembership

Represents a single cache security group and its status..

Constructors

SecurityGroupMembership

Fields

_sgmStatus :: Maybe Text

The status of the cache security group membership. The status changes whenever a cache security group is modified, or when the cache security groups assigned to a cache cluster are modified.

_sgmSecurityGroupId :: Maybe Text

The identifier of the cache security group.

data Snapshot

Represents a copy of an entire cache cluster as of the time when the snapshot was taken.

Constructors

Snapshot

Fields

_ssssssssssssssstEngineVersion :: Maybe Text

The version of the cache engine version that is used by the source cache cluster.

_ssssssssssssssstCacheNodeType :: Maybe Text

The name of the compute and memory capacity node type for the source cache cluster.

_ssssssssssssssstCacheClusterCreateTime :: Maybe ISO8601

The date and time when the source cache cluster was created.

_ssssssssssssssstAutoMinorVersionUpgrade :: Maybe Bool

For the source cache cluster, indicates whether minor version patches are applied automatically (true) or not (false).

_ssssssssssssssstCacheParameterGroupName :: Maybe Text

The cache parameter group that is associated with the source cache cluster.

_ssssssssssssssstVpcId :: Maybe Text

The Amazon Virtual Private Cloud identifier (VPC ID) of the cache subnet group for the source cache cluster.

_ssssssssssssssstSnapshotStatus :: Maybe Text

The status of the snapshot. Valid values: creating | available | restoring | deleting.

_ssssssssssssssstSnapshotWindow :: Maybe Text

The daily time range during which ElastiCache takes daily snapshots of the source cache cluster.

_ssssssssssssssstCacheClusterId :: Maybe Text

The user-supplied identifier of the source cache cluster.

_ssssssssssssssstEngine :: Maybe Text

The name of the cache engine (memcached or redis) used by the source cache cluster.

_ssssssssssssssstPreferredMaintenanceWindow :: Maybe Text

The time range (in UTC) during which weekly system maintenance can occur on the source cache cluster.

_ssssssssssssssstTopicArn :: Maybe Text

The Amazon Resource Name (ARN) for the topic used by the source cache cluster for publishing notifications.

_ssssssssssssssstNodeSnapshots :: [NodeSnapshot]

A list of the cache cluster nodes in the source cache cluster.

_ssssssssssssssstCacheSubnetGroupName :: Maybe Text

The name of the cache subnet group associated with the source cache cluster.

_ssssssssssssssstPreferredAvailabilityZone :: Maybe Text

The name of the Availability Zone in which the source cache cluster is located.

_ssssssssssssssstSnapshotRetentionLimit :: Maybe Integer

For an automatic snapshot, the number of days for which ElastiCache will retain the snapshot before deleting it. For manual snapshots, this field reflects the SnapshotRetentionLimit for the source cache cluster when the snapshot was created. This field is otherwise ignored: Manual snapshots do not expire, and can only be deleted using the DeleteSnapshot action.

_ssssssssssssssstSnapshotName :: Maybe Text

The name of a snapshot. For an automatic snapshot, the name is system-generated; for a manual snapshot, this is the user-provided name.

_ssssssssssssssstNumCacheNodes :: Maybe Integer

The number of cache nodes in the source cache cluster.

_ssssssssssssssstPort :: Maybe Integer

The port number used by each cache nodes in the source cache cluster.

_ssssssssssssssstSnapshotSource :: Maybe Text

Indicates whether the snapshot is from an automatic backup (automated) or was created manually (manual).

data Subnet

Represents the subnet associated with a cache cluster. This parameter refers to subnets defined in Amazon Virtual Private Cloud (Amazon VPC) and used with ElastiCache.

Constructors

Subnet

Fields

_ssssssssssssswSubnetIdentifier :: Maybe Text

The unique identifier for the subnet.

_ssssssssssssswSubnetAvailabilityZone :: Maybe AvailabilityZone

The Availability Zone associated with the subnet.

azName :: Iso' AvailabilityZone (Maybe Text)

availabilityZone :: Iso' (Maybe Text) AvailabilityZone

rgpmvPrimaryClusterId :: Iso' ReplicationGroupPendingModifiedValues (Maybe Text)

replicationGroupPendingModifiedValues :: Iso' (Maybe Text) ReplicationGroupPendingModifiedValues

ccSnapshotWindow :: Lens' CacheCluster (Maybe Text)

ccSnapshotRetentionLimit :: Lens' CacheCluster (Maybe Integer)

ccSecurityGroups :: Lens' CacheCluster [SecurityGroupMembership]

ccReplicationGroupId :: Lens' CacheCluster (Maybe Text)

ccPreferredMaintenanceWindow :: Lens' CacheCluster (Maybe Text)

ccPreferredAvailabilityZone :: Lens' CacheCluster (Maybe Text)

ccPendingModifiedValues :: Lens' CacheCluster (Maybe PendingModifiedValues)

ccNumCacheNodes :: Lens' CacheCluster (Maybe Integer)

ccNotificationConfiguration :: Lens' CacheCluster (Maybe NotificationConfiguration)

ccEngineVersion :: Lens' CacheCluster (Maybe Text)

ccEngine :: Lens' CacheCluster (Maybe Text)

ccConfigurationEndpoint :: Lens' CacheCluster (Maybe Endpoint)

ccClientDownloadLandingPage :: Lens' CacheCluster (Maybe Text)

ccCacheSubnetGroupName :: Lens' CacheCluster (Maybe Text)

ccCacheSecurityGroups :: Lens' CacheCluster [CacheSecurityGroupMembership]

ccCacheParameterGroup :: Lens' CacheCluster (Maybe CacheParameterGroupStatus)

ccCacheNodes :: Lens' CacheCluster [CacheNode]

ccCacheNodeType :: Lens' CacheCluster (Maybe Text)

ccCacheClusterStatus :: Lens' CacheCluster (Maybe Text)

ccCacheClusterId :: Lens' CacheCluster (Maybe Text)

ccCacheClusterCreateTime :: Lens' CacheCluster (Maybe ISO8601)

ccAutoMinorVersionUpgrade :: Lens' CacheCluster (Maybe Bool)

cevEngineVersion :: Lens' CacheEngineVersion (Maybe Text)

cevEngine :: Lens' CacheEngineVersion (Maybe Text)

cevCacheParameterGroupFamily :: Lens' CacheEngineVersion (Maybe Text)

cevCacheEngineVersionDescription :: Lens' CacheEngineVersion (Maybe Text)

cevCacheEngineDescription :: Lens' CacheEngineVersion (Maybe Text)

cnSourceCacheNodeId :: Lens' CacheNode (Maybe Text)

cnParameterGroupStatus :: Lens' CacheNode (Maybe Text)

cnEndpoint :: Lens' CacheNode (Maybe Endpoint)

cnCacheNodeStatus :: Lens' CacheNode (Maybe Text)

cnCacheNodeId :: Lens' CacheNode (Maybe Text)

cnCacheNodeCreateTime :: Lens' CacheNode (Maybe ISO8601)

cntspSource :: Lens' CacheNodeTypeSpecificParameter (Maybe Text)

cntspParameterName :: Lens' CacheNodeTypeSpecificParameter (Maybe Text)

cntspMinimumEngineVersion :: Lens' CacheNodeTypeSpecificParameter (Maybe Text)

cntspIsModifiable :: Lens' CacheNodeTypeSpecificParameter (Maybe Bool)

cntspDescription :: Lens' CacheNodeTypeSpecificParameter (Maybe Text)

cntspDataType :: Lens' CacheNodeTypeSpecificParameter (Maybe Text)

cntspCacheNodeTypeSpecificValues :: Lens' CacheNodeTypeSpecificParameter [CacheNodeTypeSpecificValue]

cntspAllowedValues :: Lens' CacheNodeTypeSpecificParameter (Maybe Text)

cntsvValue :: Lens' CacheNodeTypeSpecificValue (Maybe Text)

cntsvCacheNodeType :: Lens' CacheNodeTypeSpecificValue (Maybe Text)

cpgDescription :: Lens' CacheParameterGroup (Maybe Text)

cpgCacheParameterGroupName :: Lens' CacheParameterGroup (Maybe Text)

cpgCacheParameterGroupFamily :: Lens' CacheParameterGroup (Maybe Text)

cpgsParameterApplyStatus :: Lens' CacheParameterGroupStatus (Maybe Text)

cpgsCacheParameterGroupName :: Lens' CacheParameterGroupStatus (Maybe Text)

cpgsCacheNodeIdsToReboot :: Lens' CacheParameterGroupStatus [Text]

csgOwnerId :: Lens' CacheSecurityGroup (Maybe Text)

csgEC2SecurityGroups :: Lens' CacheSecurityGroup [EC2SecurityGroup]

csgDescription :: Lens' CacheSecurityGroup (Maybe Text)

csgCacheSecurityGroupName :: Lens' CacheSecurityGroup (Maybe Text)

csgmStatus :: Lens' CacheSecurityGroupMembership (Maybe Text)

csgmCacheSecurityGroupName :: Lens' CacheSecurityGroupMembership (Maybe Text)

csiVpcId :: Lens' CacheSubnetGroup (Maybe Text)

csiSubnets :: Lens' CacheSubnetGroup [Subnet]

csiCacheSubnetGroupName :: Lens' CacheSubnetGroup (Maybe Text)

csiCacheSubnetGroupDescription :: Lens' CacheSubnetGroup (Maybe Text)

ecsgStatus :: Lens' EC2SecurityGroup (Maybe Text)

ecsgEC2SecurityGroupOwnerId :: Lens' EC2SecurityGroup (Maybe Text)

ecsgEC2SecurityGroupName :: Lens' EC2SecurityGroup (Maybe Text)

ePort :: Lens' Endpoint (Maybe Integer)

eAddress :: Lens' Endpoint (Maybe Text)

edParameters :: Lens' EngineDefaults [Parameter]

edMarker :: Lens' EngineDefaults (Maybe Text)

edCacheParameterGroupFamily :: Lens' EngineDefaults (Maybe Text)

edCacheNodeTypeSpecificParameters :: Lens' EngineDefaults [CacheNodeTypeSpecificParameter]

euSourceType :: Lens' Event (Maybe SourceType)

euSourceIdentifier :: Lens' Event (Maybe Text)

euMessage :: Lens' Event (Maybe Text)

euDate :: Lens' Event (Maybe ISO8601)

ngStatus :: Lens' NodeGroup (Maybe Text)

ngPrimaryEndpoint :: Lens' NodeGroup (Maybe Endpoint)

ngNodeGroupMembers :: Lens' NodeGroup [NodeGroupMember]

ngNodeGroupId :: Lens' NodeGroup (Maybe Text)

ngmReadEndpoint :: Lens' NodeGroupMember (Maybe Endpoint)

ngmPreferredAvailabilityZone :: Lens' NodeGroupMember (Maybe Text)

ngmCurrentRole :: Lens' NodeGroupMember (Maybe Text)

ngmCacheNodeId :: Lens' NodeGroupMember (Maybe Text)

ngmCacheClusterId :: Lens' NodeGroupMember (Maybe Text)

nsSnapshotCreateTime :: Lens' NodeSnapshot (Maybe ISO8601)

nsCacheSize :: Lens' NodeSnapshot (Maybe Text)

nsCacheNodeId :: Lens' NodeSnapshot (Maybe Text)

nsCacheNodeCreateTime :: Lens' NodeSnapshot (Maybe ISO8601)

ncTopicStatus :: Lens' NotificationConfiguration (Maybe Text)

ncTopicArn :: Lens' NotificationConfiguration (Maybe Text)

prSource :: Lens' Parameter (Maybe Text)

prParameterValue :: Lens' Parameter (Maybe Text)

prParameterName :: Lens' Parameter (Maybe Text)

prMinimumEngineVersion :: Lens' Parameter (Maybe Text)

prIsModifiable :: Lens' Parameter (Maybe Bool)

prDescription :: Lens' Parameter (Maybe Text)

prDataType :: Lens' Parameter (Maybe Text)

prAllowedValues :: Lens' Parameter (Maybe Text)

pnvParameterValue :: Lens' ParameterNameValue (Maybe Text)

pnvParameterName :: Lens' ParameterNameValue (Maybe Text)

pmvNumCacheNodes :: Lens' PendingModifiedValues (Maybe Integer)

pmvEngineVersion :: Lens' PendingModifiedValues (Maybe Text)

pmvCacheNodeIdsToRemove :: Lens' PendingModifiedValues [Text]

rcRecurringChargeFrequency :: Lens' RecurringCharge (Maybe Text)

rcRecurringChargeAmount :: Lens' RecurringCharge (Maybe Double)

rgStatus :: Lens' ReplicationGroup (Maybe Text)

rgSnapshottingClusterId :: Lens' ReplicationGroup (Maybe Text)

rgReplicationGroupId :: Lens' ReplicationGroup (Maybe Text)

rgPendingModifiedValues :: Lens' ReplicationGroup (Maybe ReplicationGroupPendingModifiedValues)

rgNodeGroups :: Lens' ReplicationGroup [NodeGroup]

rgMemberClusters :: Lens' ReplicationGroup [Text]

rgDescription :: Lens' ReplicationGroup (Maybe Text)

rcnUsagePrice :: Lens' ReservedCacheNode (Maybe Double)

rcnState :: Lens' ReservedCacheNode (Maybe Text)

rcnStartTime :: Lens' ReservedCacheNode (Maybe ISO8601)

rcnReservedCacheNodesOfferingId :: Lens' ReservedCacheNode (Maybe Text)

rcnReservedCacheNodeId :: Lens' ReservedCacheNode (Maybe Text)

rcnRecurringCharges :: Lens' ReservedCacheNode [RecurringCharge]

rcnProductDescription :: Lens' ReservedCacheNode (Maybe Text)

rcnOfferingType :: Lens' ReservedCacheNode (Maybe Text)

rcnFixedPrice :: Lens' ReservedCacheNode (Maybe Double)

rcnDuration :: Lens' ReservedCacheNode (Maybe Integer)

rcnCacheNodeType :: Lens' ReservedCacheNode (Maybe Text)

rcnCacheNodeCount :: Lens' ReservedCacheNode (Maybe Integer)

rcnoUsagePrice :: Lens' ReservedCacheNodesOffering (Maybe Double)

rcnoReservedCacheNodesOfferingId :: Lens' ReservedCacheNodesOffering (Maybe Text)

rcnoRecurringCharges :: Lens' ReservedCacheNodesOffering [RecurringCharge]

rcnoProductDescription :: Lens' ReservedCacheNodesOffering (Maybe Text)

rcnoOfferingType :: Lens' ReservedCacheNodesOffering (Maybe Text)

rcnoFixedPrice :: Lens' ReservedCacheNodesOffering (Maybe Double)

rcnoDuration :: Lens' ReservedCacheNodesOffering (Maybe Integer)

rcnoCacheNodeType :: Lens' ReservedCacheNodesOffering (Maybe Text)

sgmStatus :: Lens' SecurityGroupMembership (Maybe Text)

sgmSecurityGroupId :: Lens' SecurityGroupMembership (Maybe Text)

ssssssssssssssstVpcId :: Lens' Snapshot (Maybe Text)

ssssssssssssssstTopicArn :: Lens' Snapshot (Maybe Text)

ssssssssssssssstSnapshotWindow :: Lens' Snapshot (Maybe Text)

ssssssssssssssstSnapshotStatus :: Lens' Snapshot (Maybe Text)

ssssssssssssssstSnapshotSource :: Lens' Snapshot (Maybe Text)

ssssssssssssssstSnapshotRetentionLimit :: Lens' Snapshot (Maybe Integer)

ssssssssssssssstSnapshotName :: Lens' Snapshot (Maybe Text)

ssssssssssssssstPreferredMaintenanceWindow :: Lens' Snapshot (Maybe Text)

ssssssssssssssstPreferredAvailabilityZone :: Lens' Snapshot (Maybe Text)

ssssssssssssssstPort :: Lens' Snapshot (Maybe Integer)

ssssssssssssssstNumCacheNodes :: Lens' Snapshot (Maybe Integer)

ssssssssssssssstNodeSnapshots :: Lens' Snapshot [NodeSnapshot]

ssssssssssssssstEngineVersion :: Lens' Snapshot (Maybe Text)

ssssssssssssssstEngine :: Lens' Snapshot (Maybe Text)

ssssssssssssssstCacheSubnetGroupName :: Lens' Snapshot (Maybe Text)

ssssssssssssssstCacheParameterGroupName :: Lens' Snapshot (Maybe Text)

ssssssssssssssstCacheNodeType :: Lens' Snapshot (Maybe Text)

ssssssssssssssstCacheClusterId :: Lens' Snapshot (Maybe Text)

ssssssssssssssstCacheClusterCreateTime :: Lens' Snapshot (Maybe ISO8601)

ssssssssssssssstAutoMinorVersionUpgrade :: Lens' Snapshot (Maybe Bool)

ssssssssssssswSubnetIdentifier :: Lens' Subnet (Maybe Text)

ssssssssssssswSubnetAvailabilityZone :: Lens' Subnet (Maybe AvailabilityZone)