Network.AWS.ElasticBeanstalk.ApplyEnvironmentManagedAction

Description

Applies a scheduled managed action immediately. A managed action can be applied only if its status is Scheduled. Get the status and action ID of a managed action with DescribeEnvironmentManagedActions.

Creating a Request

data ApplyEnvironmentManagedAction

Request to execute a scheduled managed action immediately.

See: newApplyEnvironmentManagedAction smart constructor.

Constructors

ApplyEnvironmentManagedAction'

Fields

environmentId :: Maybe Text

The environment ID of the target environment.

environmentName :: Maybe Text

The name of the target environment.

actionId :: Text

The action ID of the scheduled managed action to execute.

newApplyEnvironmentManagedAction

Arguments

:: Text
$sel:actionId:ApplyEnvironmentManagedAction'

-> ApplyEnvironmentManagedAction

Create a value of ApplyEnvironmentManagedAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:environmentId:ApplyEnvironmentManagedAction', applyEnvironmentManagedAction_environmentId - The environment ID of the target environment.

$sel:environmentName:ApplyEnvironmentManagedAction', applyEnvironmentManagedAction_environmentName - The name of the target environment.

$sel:actionId:ApplyEnvironmentManagedAction', applyEnvironmentManagedAction_actionId - The action ID of the scheduled managed action to execute.

Request Lenses

applyEnvironmentManagedAction_environmentId :: Lens' ApplyEnvironmentManagedAction (Maybe Text)

The environment ID of the target environment.

applyEnvironmentManagedAction_environmentName :: Lens' ApplyEnvironmentManagedAction (Maybe Text)

The name of the target environment.

applyEnvironmentManagedAction_actionId :: Lens' ApplyEnvironmentManagedAction Text

The action ID of the scheduled managed action to execute.

Destructuring the Response

data ApplyEnvironmentManagedActionResponse

The result message containing information about the managed action.

See: newApplyEnvironmentManagedActionResponse smart constructor.

Constructors

ApplyEnvironmentManagedActionResponse'

Fields

status :: Maybe Text

The status of the managed action.

actionType :: Maybe ActionType

The type of managed action.

actionId :: Maybe Text

The action ID of the managed action.

actionDescription :: Maybe Text

A description of the managed action.

httpStatus :: Int

The response's http status code.

newApplyEnvironmentManagedActionResponse

Arguments

:: Int
$sel:httpStatus:ApplyEnvironmentManagedActionResponse'

-> ApplyEnvironmentManagedActionResponse

Create a value of ApplyEnvironmentManagedActionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ApplyEnvironmentManagedActionResponse', applyEnvironmentManagedActionResponse_status - The status of the managed action.

$sel:actionType:ApplyEnvironmentManagedActionResponse', applyEnvironmentManagedActionResponse_actionType - The type of managed action.

$sel:actionId:ApplyEnvironmentManagedAction', applyEnvironmentManagedActionResponse_actionId - The action ID of the managed action.

$sel:actionDescription:ApplyEnvironmentManagedActionResponse', applyEnvironmentManagedActionResponse_actionDescription - A description of the managed action.

$sel:httpStatus:ApplyEnvironmentManagedActionResponse', applyEnvironmentManagedActionResponse_httpStatus - The response's http status code.

Response Lenses

applyEnvironmentManagedActionResponse_status :: Lens' ApplyEnvironmentManagedActionResponse (Maybe Text)

The status of the managed action.

applyEnvironmentManagedActionResponse_actionType :: Lens' ApplyEnvironmentManagedActionResponse (Maybe ActionType)

The type of managed action.

applyEnvironmentManagedActionResponse_actionId :: Lens' ApplyEnvironmentManagedActionResponse (Maybe Text)

The action ID of the managed action.

applyEnvironmentManagedActionResponse_actionDescription :: Lens' ApplyEnvironmentManagedActionResponse (Maybe Text)

A description of the managed action.

applyEnvironmentManagedActionResponse_httpStatus :: Lens' ApplyEnvironmentManagedActionResponse Int

The response's http status code.