Network.AWS.ElasticBeanstalk.CheckDNSAvailability

Description

Checks if the specified CNAME is available.

Creating a Request

data CheckDNSAvailability

Results message indicating whether a CNAME is available.

See: newCheckDNSAvailability smart constructor.

Constructors

CheckDNSAvailability'

Fields

cNAMEPrefix :: Text

The prefix used when this CNAME is reserved.

newCheckDNSAvailability

Arguments

:: Text
$sel:cNAMEPrefix:CheckDNSAvailability'

-> CheckDNSAvailability

Create a value of CheckDNSAvailability with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cNAMEPrefix:CheckDNSAvailability', checkDNSAvailability_cNAMEPrefix - The prefix used when this CNAME is reserved.

Request Lenses

checkDNSAvailability_cNAMEPrefix :: Lens' CheckDNSAvailability Text

The prefix used when this CNAME is reserved.

Destructuring the Response

data CheckDNSAvailabilityResponse

Indicates if the specified CNAME is available.

See: newCheckDNSAvailabilityResponse smart constructor.

Constructors

CheckDNSAvailabilityResponse'

Fields

fullyQualifiedCNAME :: Maybe Text

The fully qualified CNAME to reserve when CreateEnvironment is called with the provided prefix.

available :: Maybe Bool

Indicates if the specified CNAME is available:

true : The CNAME is available.

false : The CNAME is not available.

httpStatus :: Int

The response's http status code.

newCheckDNSAvailabilityResponse

Arguments

:: Int
$sel:httpStatus:CheckDNSAvailabilityResponse'

-> CheckDNSAvailabilityResponse

Create a value of CheckDNSAvailabilityResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fullyQualifiedCNAME:CheckDNSAvailabilityResponse', checkDNSAvailabilityResponse_fullyQualifiedCNAME - The fully qualified CNAME to reserve when CreateEnvironment is called with the provided prefix.

$sel:available:CheckDNSAvailabilityResponse', checkDNSAvailabilityResponse_available - Indicates if the specified CNAME is available:

true : The CNAME is available.

false : The CNAME is not available.

$sel:httpStatus:CheckDNSAvailabilityResponse', checkDNSAvailabilityResponse_httpStatus - The response's http status code.

Response Lenses

checkDNSAvailabilityResponse_fullyQualifiedCNAME :: Lens' CheckDNSAvailabilityResponse (Maybe Text)

The fully qualified CNAME to reserve when CreateEnvironment is called with the provided prefix.

checkDNSAvailabilityResponse_available :: Lens' CheckDNSAvailabilityResponse (Maybe Bool)

Indicates if the specified CNAME is available:

true : The CNAME is available.

false : The CNAME is not available.

checkDNSAvailabilityResponse_httpStatus :: Lens' CheckDNSAvailabilityResponse Int

The response's http status code.