Network.AWS.ElasticBeanstalk.ComposeEnvironments

Description

Create or update a group of environments that each run a separate component of a single application. Takes a list of version labels that specify application source bundles for each of the environments to create or update. The name of each environment and other required information must be included in the source bundles in an environment manifest named env.yaml. See Compose Environments for details.

Creating a Request

data ComposeEnvironments

Request to create or update a group of environments.

See: newComposeEnvironments smart constructor.

Constructors

ComposeEnvironments'

Fields

groupName :: Maybe Text

The name of the group to which the target environments belong. Specify a group name only if the environment name defined in each target environment's manifest ends with a + (plus) character. See Environment Manifest (env.yaml) for details.

versionLabels :: Maybe [Text]

A list of version labels, specifying one or more application source bundles that belong to the target application. Each source bundle must include an environment manifest that specifies the name of the environment and the name of the solution stack to use, and optionally can specify environment links to create.

applicationName :: Maybe Text

The name of the application to which the specified source bundles belong.

newComposeEnvironments :: ComposeEnvironments

Create a value of ComposeEnvironments with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groupName:ComposeEnvironments', composeEnvironments_groupName - The name of the group to which the target environments belong. Specify a group name only if the environment name defined in each target environment's manifest ends with a + (plus) character. See Environment Manifest (env.yaml) for details.

$sel:versionLabels:ComposeEnvironments', composeEnvironments_versionLabels - A list of version labels, specifying one or more application source bundles that belong to the target application. Each source bundle must include an environment manifest that specifies the name of the environment and the name of the solution stack to use, and optionally can specify environment links to create.

$sel:applicationName:ComposeEnvironments', composeEnvironments_applicationName - The name of the application to which the specified source bundles belong.

Request Lenses

composeEnvironments_groupName :: Lens' ComposeEnvironments (Maybe Text)

The name of the group to which the target environments belong. Specify a group name only if the environment name defined in each target environment's manifest ends with a + (plus) character. See Environment Manifest (env.yaml) for details.

composeEnvironments_versionLabels :: Lens' ComposeEnvironments (Maybe [Text])

A list of version labels, specifying one or more application source bundles that belong to the target application. Each source bundle must include an environment manifest that specifies the name of the environment and the name of the solution stack to use, and optionally can specify environment links to create.

composeEnvironments_applicationName :: Lens' ComposeEnvironments (Maybe Text)

The name of the application to which the specified source bundles belong.

Destructuring the Response

data EnvironmentDescriptionsMessage

Result message containing a list of environment descriptions.

See: newEnvironmentDescriptionsMessage smart constructor.

Constructors

EnvironmentDescriptionsMessage'

Fields

nextToken :: Maybe Text

In a paginated request, the token that you can pass in a subsequent request to get the next response page.

environments :: Maybe [EnvironmentDescription]

Returns an EnvironmentDescription list.

newEnvironmentDescriptionsMessage :: EnvironmentDescriptionsMessage

Create a value of EnvironmentDescriptionsMessage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:EnvironmentDescriptionsMessage', environmentDescriptionsMessage_nextToken - In a paginated request, the token that you can pass in a subsequent request to get the next response page.

$sel:environments:EnvironmentDescriptionsMessage', environmentDescriptionsMessage_environments - Returns an EnvironmentDescription list.

Response Lenses

environmentDescriptionsMessage_nextToken :: Lens' EnvironmentDescriptionsMessage (Maybe Text)

In a paginated request, the token that you can pass in a subsequent request to get the next response page.

environmentDescriptionsMessage_environments :: Lens' EnvironmentDescriptionsMessage (Maybe [EnvironmentDescription])

Returns an EnvironmentDescription list.