Network.AWS.ElasticBeanstalk.CreateApplication

Description

Creates an application that has one configuration template named default and no application versions.

Creating a Request

data CreateApplication

Request to create an application.

See: newCreateApplication smart constructor.

Constructors

CreateApplication'

Fields

tags :: Maybe [Tag]

Specifies the tags applied to the application.

Elastic Beanstalk applies these tags only to the application. Environments that you create in the application don't inherit the tags.

resourceLifecycleConfig :: Maybe ApplicationResourceLifecycleConfig

Specifies an application resource lifecycle configuration to prevent your application from accumulating too many versions.

description :: Maybe Text

Your description of the application.

applicationName :: Text

The name of the application. Must be unique within your account.

newCreateApplication

Arguments

:: Text
$sel:applicationName:CreateApplication'

-> CreateApplication

Create a value of CreateApplication with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateApplication', createApplication_tags - Specifies the tags applied to the application.

Elastic Beanstalk applies these tags only to the application. Environments that you create in the application don't inherit the tags.

$sel:resourceLifecycleConfig:CreateApplication', createApplication_resourceLifecycleConfig - Specifies an application resource lifecycle configuration to prevent your application from accumulating too many versions.

$sel:description:CreateApplication', createApplication_description - Your description of the application.

$sel:applicationName:CreateApplication', createApplication_applicationName - The name of the application. Must be unique within your account.

Request Lenses

createApplication_tags :: Lens' CreateApplication (Maybe [Tag])

Specifies the tags applied to the application.

Elastic Beanstalk applies these tags only to the application. Environments that you create in the application don't inherit the tags.

createApplication_resourceLifecycleConfig :: Lens' CreateApplication (Maybe ApplicationResourceLifecycleConfig)

Specifies an application resource lifecycle configuration to prevent your application from accumulating too many versions.

createApplication_description :: Lens' CreateApplication (Maybe Text)

Your description of the application.

createApplication_applicationName :: Lens' CreateApplication Text

The name of the application. Must be unique within your account.

Destructuring the Response

data ApplicationDescriptionMessage

Result message containing a single description of an application.

See: newApplicationDescriptionMessage smart constructor.

Constructors

ApplicationDescriptionMessage'

Fields

application :: Maybe ApplicationDescription

The ApplicationDescription of the application.

newApplicationDescriptionMessage :: ApplicationDescriptionMessage

Create a value of ApplicationDescriptionMessage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:application:ApplicationDescriptionMessage', applicationDescriptionMessage_application - The ApplicationDescription of the application.

Response Lenses

applicationDescriptionMessage_application :: Lens' ApplicationDescriptionMessage (Maybe ApplicationDescription)

The ApplicationDescription of the application.