Network.AWS.ElasticBeanstalk.CreateApplicationVersion

Description

Creates an application version for the specified application. You can create an application version from a source bundle in Amazon S3, a commit in AWS CodeCommit, or the output of an AWS CodeBuild build as follows:

Specify a commit in an AWS CodeCommit repository with SourceBuildInformation.

Specify a build in an AWS CodeBuild with SourceBuildInformation and BuildConfiguration.

Specify a source bundle in S3 with SourceBundle

Omit both SourceBuildInformation and SourceBundle to use the default sample application.

After you create an application version with a specified Amazon S3 bucket and key location, you can't change that Amazon S3 location. If you change the Amazon S3 location, you receive an exception when you attempt to launch an environment from the application version.

Creating a Request

data CreateApplicationVersion

See: newCreateApplicationVersion smart constructor.

Constructors

CreateApplicationVersion'

Fields

process :: Maybe Bool

Pre-processes and validates the environment manifest (env.yaml) and configuration files (*.config files in the .ebextensions folder) in the source bundle. Validating configuration files can identify issues prior to deploying the application version to an environment.

You must turn processing on for application versions that you create using AWS CodeBuild or AWS CodeCommit. For application versions built from a source bundle in Amazon S3, processing is optional.

The Process option validates Elastic Beanstalk configuration files. It doesn't validate your application's configuration files, like proxy server or Docker configuration.

sourceBundle :: Maybe S3Location

The Amazon S3 bucket and key that identify the location of the source bundle for this version.

The Amazon S3 bucket must be in the same region as the environment.

Specify a source bundle in S3 or a commit in an AWS CodeCommit repository (with SourceBuildInformation), but not both. If neither SourceBundle nor SourceBuildInformation are provided, Elastic Beanstalk uses a sample application.

autoCreateApplication :: Maybe Bool

Set to true to create an application with the specified name if it doesn't already exist.

sourceBuildInformation :: Maybe SourceBuildInformation

Specify a commit in an AWS CodeCommit Git repository to use as the source code for the application version.

description :: Maybe Text

A description of this application version.

buildConfiguration :: Maybe BuildConfiguration

Settings for an AWS CodeBuild build.

tags :: Maybe [Tag]

Specifies the tags applied to the application version.

Elastic Beanstalk applies these tags only to the application version. Environments that use the application version don't inherit the tags.

applicationName :: Text

The name of the application. If no application is found with this name, and AutoCreateApplication is false, returns an InvalidParameterValue error.

versionLabel :: Text

A label identifying this version.

Constraint: Must be unique per application. If an application version already exists with this label for the specified application, AWS Elastic Beanstalk returns an InvalidParameterValue error.

newCreateApplicationVersion

Arguments

:: Text
$sel:applicationName:CreateApplicationVersion'

-> Text
$sel:versionLabel:CreateApplicationVersion'

-> CreateApplicationVersion

Create a value of CreateApplicationVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:process:CreateApplicationVersion', createApplicationVersion_process - Pre-processes and validates the environment manifest (env.yaml) and configuration files (*.config files in the .ebextensions folder) in the source bundle. Validating configuration files can identify issues prior to deploying the application version to an environment.

You must turn processing on for application versions that you create using AWS CodeBuild or AWS CodeCommit. For application versions built from a source bundle in Amazon S3, processing is optional.

The Process option validates Elastic Beanstalk configuration files. It doesn't validate your application's configuration files, like proxy server or Docker configuration.

$sel:sourceBundle:CreateApplicationVersion', createApplicationVersion_sourceBundle - The Amazon S3 bucket and key that identify the location of the source bundle for this version.

The Amazon S3 bucket must be in the same region as the environment.

Specify a source bundle in S3 or a commit in an AWS CodeCommit repository (with SourceBuildInformation), but not both. If neither SourceBundle nor SourceBuildInformation are provided, Elastic Beanstalk uses a sample application.

$sel:autoCreateApplication:CreateApplicationVersion', createApplicationVersion_autoCreateApplication - Set to true to create an application with the specified name if it doesn't already exist.

$sel:sourceBuildInformation:CreateApplicationVersion', createApplicationVersion_sourceBuildInformation - Specify a commit in an AWS CodeCommit Git repository to use as the source code for the application version.

$sel:description:CreateApplicationVersion', createApplicationVersion_description - A description of this application version.

$sel:buildConfiguration:CreateApplicationVersion', createApplicationVersion_buildConfiguration - Settings for an AWS CodeBuild build.

$sel:tags:CreateApplicationVersion', createApplicationVersion_tags - Specifies the tags applied to the application version.

Elastic Beanstalk applies these tags only to the application version. Environments that use the application version don't inherit the tags.

$sel:applicationName:CreateApplicationVersion', createApplicationVersion_applicationName - The name of the application. If no application is found with this name, and AutoCreateApplication is false, returns an InvalidParameterValue error.

$sel:versionLabel:CreateApplicationVersion', createApplicationVersion_versionLabel - A label identifying this version.

Constraint: Must be unique per application. If an application version already exists with this label for the specified application, AWS Elastic Beanstalk returns an InvalidParameterValue error.

Request Lenses

createApplicationVersion_process :: Lens' CreateApplicationVersion (Maybe Bool)

Pre-processes and validates the environment manifest (env.yaml) and configuration files (*.config files in the .ebextensions folder) in the source bundle. Validating configuration files can identify issues prior to deploying the application version to an environment.

You must turn processing on for application versions that you create using AWS CodeBuild or AWS CodeCommit. For application versions built from a source bundle in Amazon S3, processing is optional.

The Process option validates Elastic Beanstalk configuration files. It doesn't validate your application's configuration files, like proxy server or Docker configuration.

createApplicationVersion_sourceBundle :: Lens' CreateApplicationVersion (Maybe S3Location)

The Amazon S3 bucket and key that identify the location of the source bundle for this version.

The Amazon S3 bucket must be in the same region as the environment.

Specify a source bundle in S3 or a commit in an AWS CodeCommit repository (with SourceBuildInformation), but not both. If neither SourceBundle nor SourceBuildInformation are provided, Elastic Beanstalk uses a sample application.

createApplicationVersion_autoCreateApplication :: Lens' CreateApplicationVersion (Maybe Bool)

Set to true to create an application with the specified name if it doesn't already exist.

createApplicationVersion_sourceBuildInformation :: Lens' CreateApplicationVersion (Maybe SourceBuildInformation)

Specify a commit in an AWS CodeCommit Git repository to use as the source code for the application version.

createApplicationVersion_description :: Lens' CreateApplicationVersion (Maybe Text)

A description of this application version.

createApplicationVersion_buildConfiguration :: Lens' CreateApplicationVersion (Maybe BuildConfiguration)

Settings for an AWS CodeBuild build.

createApplicationVersion_tags :: Lens' CreateApplicationVersion (Maybe [Tag])

Specifies the tags applied to the application version.

Elastic Beanstalk applies these tags only to the application version. Environments that use the application version don't inherit the tags.

createApplicationVersion_applicationName :: Lens' CreateApplicationVersion Text

The name of the application. If no application is found with this name, and AutoCreateApplication is false, returns an InvalidParameterValue error.

createApplicationVersion_versionLabel :: Lens' CreateApplicationVersion Text

A label identifying this version.

Constraint: Must be unique per application. If an application version already exists with this label for the specified application, AWS Elastic Beanstalk returns an InvalidParameterValue error.

Destructuring the Response

data ApplicationVersionDescriptionMessage

Result message wrapping a single description of an application version.

See: newApplicationVersionDescriptionMessage smart constructor.

Constructors

ApplicationVersionDescriptionMessage'

Fields

applicationVersion :: Maybe ApplicationVersionDescription

The ApplicationVersionDescription of the application version.

newApplicationVersionDescriptionMessage :: ApplicationVersionDescriptionMessage

Create a value of ApplicationVersionDescriptionMessage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationVersion:ApplicationVersionDescriptionMessage', applicationVersionDescriptionMessage_applicationVersion - The ApplicationVersionDescription of the application version.

Response Lenses

applicationVersionDescriptionMessage_applicationVersion :: Lens' ApplicationVersionDescriptionMessage (Maybe ApplicationVersionDescription)

The ApplicationVersionDescription of the application version.