Network.AWS.ElasticBeanstalk.CreateConfigurationTemplate

Description

Creates an AWS Elastic Beanstalk configuration template, associated with a specific Elastic Beanstalk application. You define application configuration settings in a configuration template. You can then use the configuration template to deploy different versions of the application with the same configuration settings.

Templates aren't associated with any environment. The EnvironmentName response element is always null.

Related Topics

DescribeConfigurationOptions

DescribeConfigurationSettings

ListAvailableSolutionStacks

Creating a Request

data CreateConfigurationTemplate

Request to create a configuration template.

See: newCreateConfigurationTemplate smart constructor.

Constructors

CreateConfigurationTemplate'

Fields

optionSettings :: Maybe [ConfigurationOptionSetting]

Option values for the Elastic Beanstalk configuration, such as the instance type. If specified, these values override the values obtained from the solution stack or the source configuration template. For a complete list of Elastic Beanstalk configuration options, see Option Values in the AWS Elastic Beanstalk Developer Guide.

platformArn :: Maybe Text

The Amazon Resource Name (ARN) of the custom platform. For more information, see Custom Platforms in the AWS Elastic Beanstalk Developer Guide.

If you specify PlatformArn, then don't specify SolutionStackName.

sourceConfiguration :: Maybe SourceConfiguration

An Elastic Beanstalk configuration template to base this one on. If specified, Elastic Beanstalk uses the configuration values from the specified configuration template to create a new configuration.

Values specified in OptionSettings override any values obtained from the SourceConfiguration.

You must specify SourceConfiguration if you don't specify PlatformArn, EnvironmentId, or SolutionStackName.

Constraint: If both solution stack name and source configuration are specified, the solution stack of the source configuration template must match the specified solution stack name.

solutionStackName :: Maybe Text

The name of an Elastic Beanstalk solution stack (platform version) that this configuration uses. For example, 64bit Amazon Linux 2013.09 running Tomcat 7 Java 7. A solution stack specifies the operating system, runtime, and application server for a configuration template. It also determines the set of configuration options as well as the possible and default values. For more information, see Supported Platforms in the AWS Elastic Beanstalk Developer Guide.

You must specify SolutionStackName if you don't specify PlatformArn, EnvironmentId, or SourceConfiguration.

Use the ListAvailableSolutionStacks API to obtain a list of available solution stacks.

environmentId :: Maybe Text

The ID of an environment whose settings you want to use to create the configuration template. You must specify EnvironmentId if you don't specify PlatformArn, SolutionStackName, or SourceConfiguration.

description :: Maybe Text

An optional description for this configuration.

tags :: Maybe [Tag]

Specifies the tags applied to the configuration template.

applicationName :: Text

The name of the Elastic Beanstalk application to associate with this configuration template.

templateName :: Text

The name of the configuration template.

Constraint: This name must be unique per application.

newCreateConfigurationTemplate

Arguments

:: Text
$sel:applicationName:CreateConfigurationTemplate'

-> Text
$sel:templateName:CreateConfigurationTemplate'

-> CreateConfigurationTemplate

Create a value of CreateConfigurationTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:optionSettings:CreateConfigurationTemplate', createConfigurationTemplate_optionSettings - Option values for the Elastic Beanstalk configuration, such as the instance type. If specified, these values override the values obtained from the solution stack or the source configuration template. For a complete list of Elastic Beanstalk configuration options, see Option Values in the AWS Elastic Beanstalk Developer Guide.

$sel:platformArn:CreateConfigurationTemplate', createConfigurationTemplate_platformArn - The Amazon Resource Name (ARN) of the custom platform. For more information, see Custom Platforms in the AWS Elastic Beanstalk Developer Guide.

If you specify PlatformArn, then don't specify SolutionStackName.

$sel:sourceConfiguration:CreateConfigurationTemplate', createConfigurationTemplate_sourceConfiguration - An Elastic Beanstalk configuration template to base this one on. If specified, Elastic Beanstalk uses the configuration values from the specified configuration template to create a new configuration.

Values specified in OptionSettings override any values obtained from the SourceConfiguration.

You must specify SourceConfiguration if you don't specify PlatformArn, EnvironmentId, or SolutionStackName.

Constraint: If both solution stack name and source configuration are specified, the solution stack of the source configuration template must match the specified solution stack name.

$sel:solutionStackName:CreateConfigurationTemplate', createConfigurationTemplate_solutionStackName - The name of an Elastic Beanstalk solution stack (platform version) that this configuration uses. For example, 64bit Amazon Linux 2013.09 running Tomcat 7 Java 7. A solution stack specifies the operating system, runtime, and application server for a configuration template. It also determines the set of configuration options as well as the possible and default values. For more information, see Supported Platforms in the AWS Elastic Beanstalk Developer Guide.

You must specify SolutionStackName if you don't specify PlatformArn, EnvironmentId, or SourceConfiguration.

Use the ListAvailableSolutionStacks API to obtain a list of available solution stacks.

$sel:environmentId:CreateConfigurationTemplate', createConfigurationTemplate_environmentId - The ID of an environment whose settings you want to use to create the configuration template. You must specify EnvironmentId if you don't specify PlatformArn, SolutionStackName, or SourceConfiguration.

$sel:description:CreateConfigurationTemplate', createConfigurationTemplate_description - An optional description for this configuration.

$sel:tags:CreateConfigurationTemplate', createConfigurationTemplate_tags - Specifies the tags applied to the configuration template.

$sel:applicationName:CreateConfigurationTemplate', createConfigurationTemplate_applicationName - The name of the Elastic Beanstalk application to associate with this configuration template.

$sel:templateName:CreateConfigurationTemplate', createConfigurationTemplate_templateName - The name of the configuration template.

Constraint: This name must be unique per application.

Request Lenses

createConfigurationTemplate_optionSettings :: Lens' CreateConfigurationTemplate (Maybe [ConfigurationOptionSetting])

Option values for the Elastic Beanstalk configuration, such as the instance type. If specified, these values override the values obtained from the solution stack or the source configuration template. For a complete list of Elastic Beanstalk configuration options, see Option Values in the AWS Elastic Beanstalk Developer Guide.

createConfigurationTemplate_platformArn :: Lens' CreateConfigurationTemplate (Maybe Text)

The Amazon Resource Name (ARN) of the custom platform. For more information, see Custom Platforms in the AWS Elastic Beanstalk Developer Guide.

If you specify PlatformArn, then don't specify SolutionStackName.

createConfigurationTemplate_sourceConfiguration :: Lens' CreateConfigurationTemplate (Maybe SourceConfiguration)

An Elastic Beanstalk configuration template to base this one on. If specified, Elastic Beanstalk uses the configuration values from the specified configuration template to create a new configuration.

Values specified in OptionSettings override any values obtained from the SourceConfiguration.

You must specify SourceConfiguration if you don't specify PlatformArn, EnvironmentId, or SolutionStackName.

Constraint: If both solution stack name and source configuration are specified, the solution stack of the source configuration template must match the specified solution stack name.

createConfigurationTemplate_solutionStackName :: Lens' CreateConfigurationTemplate (Maybe Text)

The name of an Elastic Beanstalk solution stack (platform version) that this configuration uses. For example, 64bit Amazon Linux 2013.09 running Tomcat 7 Java 7. A solution stack specifies the operating system, runtime, and application server for a configuration template. It also determines the set of configuration options as well as the possible and default values. For more information, see Supported Platforms in the AWS Elastic Beanstalk Developer Guide.

You must specify SolutionStackName if you don't specify PlatformArn, EnvironmentId, or SourceConfiguration.

Use the ListAvailableSolutionStacks API to obtain a list of available solution stacks.

createConfigurationTemplate_environmentId :: Lens' CreateConfigurationTemplate (Maybe Text)

The ID of an environment whose settings you want to use to create the configuration template. You must specify EnvironmentId if you don't specify PlatformArn, SolutionStackName, or SourceConfiguration.

createConfigurationTemplate_description :: Lens' CreateConfigurationTemplate (Maybe Text)

An optional description for this configuration.

createConfigurationTemplate_tags :: Lens' CreateConfigurationTemplate (Maybe [Tag])

Specifies the tags applied to the configuration template.

createConfigurationTemplate_applicationName :: Lens' CreateConfigurationTemplate Text

The name of the Elastic Beanstalk application to associate with this configuration template.

createConfigurationTemplate_templateName :: Lens' CreateConfigurationTemplate Text

The name of the configuration template.

Constraint: This name must be unique per application.

Destructuring the Response

data ConfigurationSettingsDescription

Describes the settings for a configuration set.

See: newConfigurationSettingsDescription smart constructor.

Constructors

ConfigurationSettingsDescription'

Fields

templateName :: Maybe Text

If not null, the name of the configuration template for this configuration set.

optionSettings :: Maybe [ConfigurationOptionSetting]

A list of the configuration options and their values in this configuration set.

dateUpdated :: Maybe ISO8601

The date (in UTC time) when this configuration set was last modified.

dateCreated :: Maybe ISO8601

The date (in UTC time) when this configuration set was created.

platformArn :: Maybe Text

The ARN of the platform version.

environmentName :: Maybe Text

If not null, the name of the environment for this configuration set.

applicationName :: Maybe Text

The name of the application associated with this configuration set.

deploymentStatus :: Maybe ConfigurationDeploymentStatus

If this configuration set is associated with an environment, the DeploymentStatus parameter indicates the deployment status of this configuration set:

null: This configuration is not associated with a running environment.

pending: This is a draft configuration that is not deployed to the associated environment but is in the process of deploying.

deployed: This is the configuration that is currently deployed to the associated running environment.

failed: This is a draft configuration that failed to successfully deploy.

solutionStackName :: Maybe Text

The name of the solution stack this configuration set uses.

description :: Maybe Text

Describes this configuration set.

newConfigurationSettingsDescription :: ConfigurationSettingsDescription

Create a value of ConfigurationSettingsDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:templateName:ConfigurationSettingsDescription', configurationSettingsDescription_templateName - If not null, the name of the configuration template for this configuration set.

$sel:optionSettings:ConfigurationSettingsDescription', configurationSettingsDescription_optionSettings - A list of the configuration options and their values in this configuration set.

$sel:dateUpdated:ConfigurationSettingsDescription', configurationSettingsDescription_dateUpdated - The date (in UTC time) when this configuration set was last modified.

$sel:dateCreated:ConfigurationSettingsDescription', configurationSettingsDescription_dateCreated - The date (in UTC time) when this configuration set was created.

$sel:platformArn:ConfigurationSettingsDescription', configurationSettingsDescription_platformArn - The ARN of the platform version.

$sel:environmentName:ConfigurationSettingsDescription', configurationSettingsDescription_environmentName - If not null, the name of the environment for this configuration set.

$sel:applicationName:ConfigurationSettingsDescription', configurationSettingsDescription_applicationName - The name of the application associated with this configuration set.

$sel:deploymentStatus:ConfigurationSettingsDescription', configurationSettingsDescription_deploymentStatus - If this configuration set is associated with an environment, the DeploymentStatus parameter indicates the deployment status of this configuration set:

null: This configuration is not associated with a running environment.

pending: This is a draft configuration that is not deployed to the associated environment but is in the process of deploying.

deployed: This is the configuration that is currently deployed to the associated running environment.

failed: This is a draft configuration that failed to successfully deploy.

$sel:solutionStackName:ConfigurationSettingsDescription', configurationSettingsDescription_solutionStackName - The name of the solution stack this configuration set uses.

$sel:description:ConfigurationSettingsDescription', configurationSettingsDescription_description - Describes this configuration set.

Response Lenses

configurationSettingsDescription_templateName :: Lens' ConfigurationSettingsDescription (Maybe Text)

If not null, the name of the configuration template for this configuration set.

configurationSettingsDescription_optionSettings :: Lens' ConfigurationSettingsDescription (Maybe [ConfigurationOptionSetting])

A list of the configuration options and their values in this configuration set.

configurationSettingsDescription_dateUpdated :: Lens' ConfigurationSettingsDescription (Maybe UTCTime)

The date (in UTC time) when this configuration set was last modified.

configurationSettingsDescription_dateCreated :: Lens' ConfigurationSettingsDescription (Maybe UTCTime)

The date (in UTC time) when this configuration set was created.

configurationSettingsDescription_platformArn :: Lens' ConfigurationSettingsDescription (Maybe Text)

The ARN of the platform version.

configurationSettingsDescription_environmentName :: Lens' ConfigurationSettingsDescription (Maybe Text)

If not null, the name of the environment for this configuration set.

configurationSettingsDescription_applicationName :: Lens' ConfigurationSettingsDescription (Maybe Text)

The name of the application associated with this configuration set.

configurationSettingsDescription_deploymentStatus :: Lens' ConfigurationSettingsDescription (Maybe ConfigurationDeploymentStatus)

If this configuration set is associated with an environment, the DeploymentStatus parameter indicates the deployment status of this configuration set:

null: This configuration is not associated with a running environment.

pending: This is a draft configuration that is not deployed to the associated environment but is in the process of deploying.

deployed: This is the configuration that is currently deployed to the associated running environment.

failed: This is a draft configuration that failed to successfully deploy.

configurationSettingsDescription_solutionStackName :: Lens' ConfigurationSettingsDescription (Maybe Text)

The name of the solution stack this configuration set uses.

configurationSettingsDescription_description :: Lens' ConfigurationSettingsDescription (Maybe Text)

Describes this configuration set.