Network.AWS.ElasticBeanstalk.DescribeApplications

Description

Returns the descriptions of existing applications.

Creating a Request

data DescribeApplications

Request to describe one or more applications.

See: newDescribeApplications smart constructor.

Constructors

DescribeApplications'

Fields

applicationNames :: Maybe [Text]

If specified, AWS Elastic Beanstalk restricts the returned descriptions to only include those with the specified names.

newDescribeApplications :: DescribeApplications

Create a value of DescribeApplications with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationNames:DescribeApplications', describeApplications_applicationNames - If specified, AWS Elastic Beanstalk restricts the returned descriptions to only include those with the specified names.

Request Lenses

describeApplications_applicationNames :: Lens' DescribeApplications (Maybe [Text])

If specified, AWS Elastic Beanstalk restricts the returned descriptions to only include those with the specified names.

Destructuring the Response

data DescribeApplicationsResponse

Result message containing a list of application descriptions.

See: newDescribeApplicationsResponse smart constructor.

Constructors

DescribeApplicationsResponse'

Fields

applications :: Maybe [ApplicationDescription]

This parameter contains a list of ApplicationDescription.

httpStatus :: Int

The response's http status code.

newDescribeApplicationsResponse

Arguments

:: Int
$sel:httpStatus:DescribeApplicationsResponse'

-> DescribeApplicationsResponse

Create a value of DescribeApplicationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applications:DescribeApplicationsResponse', describeApplicationsResponse_applications - This parameter contains a list of ApplicationDescription.

$sel:httpStatus:DescribeApplicationsResponse', describeApplicationsResponse_httpStatus - The response's http status code.

Response Lenses

describeApplicationsResponse_applications :: Lens' DescribeApplicationsResponse (Maybe [ApplicationDescription])

This parameter contains a list of ApplicationDescription.

describeApplicationsResponse_httpStatus :: Lens' DescribeApplicationsResponse Int

The response's http status code.