Network.AWS.ElasticBeanstalk.DescribeConfigurationSettings

Description

Returns a description of the settings for the specified configuration set, that is, either a configuration template or the configuration set associated with a running environment.

When describing the settings for the configuration set associated with a running environment, it is possible to receive two sets of setting descriptions. One is the deployed configuration set, and the other is a draft configuration of an environment that is either in the process of deployment or that failed to deploy.

Related Topics

DeleteEnvironmentConfiguration

Creating a Request

data DescribeConfigurationSettings

Result message containing all of the configuration settings for a specified solution stack or configuration template.

See: newDescribeConfigurationSettings smart constructor.

Constructors

DescribeConfigurationSettings'

Fields

templateName :: Maybe Text

The name of the configuration template to describe.

Conditional: You must specify either this parameter or an EnvironmentName, but not both. If you specify both, AWS Elastic Beanstalk returns an InvalidParameterCombination error. If you do not specify either, AWS Elastic Beanstalk returns a MissingRequiredParameter error.

environmentName :: Maybe Text

The name of the environment to describe.

Condition: You must specify either this or a TemplateName, but not both. If you specify both, AWS Elastic Beanstalk returns an InvalidParameterCombination error. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

applicationName :: Text

The application for the environment or configuration template.

newDescribeConfigurationSettings

Arguments

:: Text
$sel:applicationName:DescribeConfigurationSettings'

-> DescribeConfigurationSettings

Create a value of DescribeConfigurationSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:templateName:DescribeConfigurationSettings', describeConfigurationSettings_templateName - The name of the configuration template to describe.

Conditional: You must specify either this parameter or an EnvironmentName, but not both. If you specify both, AWS Elastic Beanstalk returns an InvalidParameterCombination error. If you do not specify either, AWS Elastic Beanstalk returns a MissingRequiredParameter error.

$sel:environmentName:DescribeConfigurationSettings', describeConfigurationSettings_environmentName - The name of the environment to describe.

Condition: You must specify either this or a TemplateName, but not both. If you specify both, AWS Elastic Beanstalk returns an InvalidParameterCombination error. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

$sel:applicationName:DescribeConfigurationSettings', describeConfigurationSettings_applicationName - The application for the environment or configuration template.

Request Lenses

describeConfigurationSettings_templateName :: Lens' DescribeConfigurationSettings (Maybe Text)

The name of the configuration template to describe.

Conditional: You must specify either this parameter or an EnvironmentName, but not both. If you specify both, AWS Elastic Beanstalk returns an InvalidParameterCombination error. If you do not specify either, AWS Elastic Beanstalk returns a MissingRequiredParameter error.

describeConfigurationSettings_environmentName :: Lens' DescribeConfigurationSettings (Maybe Text)

The name of the environment to describe.

Condition: You must specify either this or a TemplateName, but not both. If you specify both, AWS Elastic Beanstalk returns an InvalidParameterCombination error. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

describeConfigurationSettings_applicationName :: Lens' DescribeConfigurationSettings Text

The application for the environment or configuration template.

Destructuring the Response

data DescribeConfigurationSettingsResponse

The results from a request to change the configuration settings of an environment.

See: newDescribeConfigurationSettingsResponse smart constructor.

Constructors

DescribeConfigurationSettingsResponse'

Fields

configurationSettings :: Maybe [ConfigurationSettingsDescription]

A list of ConfigurationSettingsDescription.

httpStatus :: Int

The response's http status code.

newDescribeConfigurationSettingsResponse

Arguments

:: Int
$sel:httpStatus:DescribeConfigurationSettingsResponse'

-> DescribeConfigurationSettingsResponse

Create a value of DescribeConfigurationSettingsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configurationSettings:DescribeConfigurationSettingsResponse', describeConfigurationSettingsResponse_configurationSettings - A list of ConfigurationSettingsDescription.

$sel:httpStatus:DescribeConfigurationSettingsResponse', describeConfigurationSettingsResponse_httpStatus - The response's http status code.

Response Lenses

describeConfigurationSettingsResponse_configurationSettings :: Lens' DescribeConfigurationSettingsResponse (Maybe [ConfigurationSettingsDescription])

A list of ConfigurationSettingsDescription.

describeConfigurationSettingsResponse_httpStatus :: Lens' DescribeConfigurationSettingsResponse Int

The response's http status code.