Network.AWS.ElasticBeanstalk.DescribeEnvironmentHealth

Description

Returns information about the overall health of the specified environment. The DescribeEnvironmentHealth operation is only available with AWS Elastic Beanstalk Enhanced Health.

Creating a Request

data DescribeEnvironmentHealth

See the example below to learn how to create a request body.

See: newDescribeEnvironmentHealth smart constructor.

Constructors

DescribeEnvironmentHealth'

Fields

environmentId :: Maybe Text

Specify the environment by ID.

You must specify either this or an EnvironmentName, or both.

environmentName :: Maybe Text

Specify the environment by name.

You must specify either this or an EnvironmentName, or both.

attributeNames :: Maybe [EnvironmentHealthAttribute]

Specify the response elements to return. To retrieve all attributes, set to All. If no attribute names are specified, returns the name of the environment.

newDescribeEnvironmentHealth :: DescribeEnvironmentHealth

Create a value of DescribeEnvironmentHealth with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:environmentId:DescribeEnvironmentHealth', describeEnvironmentHealth_environmentId - Specify the environment by ID.

You must specify either this or an EnvironmentName, or both.

$sel:environmentName:DescribeEnvironmentHealth', describeEnvironmentHealth_environmentName - Specify the environment by name.

You must specify either this or an EnvironmentName, or both.

$sel:attributeNames:DescribeEnvironmentHealth', describeEnvironmentHealth_attributeNames - Specify the response elements to return. To retrieve all attributes, set to All. If no attribute names are specified, returns the name of the environment.

Request Lenses

describeEnvironmentHealth_environmentId :: Lens' DescribeEnvironmentHealth (Maybe Text)

Specify the environment by ID.

You must specify either this or an EnvironmentName, or both.

describeEnvironmentHealth_environmentName :: Lens' DescribeEnvironmentHealth (Maybe Text)

Specify the environment by name.

You must specify either this or an EnvironmentName, or both.

describeEnvironmentHealth_attributeNames :: Lens' DescribeEnvironmentHealth (Maybe [EnvironmentHealthAttribute])

Specify the response elements to return. To retrieve all attributes, set to All. If no attribute names are specified, returns the name of the environment.

Destructuring the Response

data DescribeEnvironmentHealthResponse

Health details for an AWS Elastic Beanstalk environment.

See: newDescribeEnvironmentHealthResponse smart constructor.

Constructors

DescribeEnvironmentHealthResponse'

Fields

status :: Maybe EnvironmentHealth

The environment's operational status. Ready, Launching, Updating, Terminating, or Terminated.

refreshedAt :: Maybe ISO8601

The date and time that the health information was retrieved.

color :: Maybe Text

The health color of the environment.

causes :: Maybe [Text]

Descriptions of the data that contributed to the environment's current health status.

environmentName :: Maybe Text

The environment's name.

instancesHealth :: Maybe InstanceHealthSummary

Summary health information for the instances in the environment.

healthStatus :: Maybe Text

The health status of the environment. For example, Ok.

applicationMetrics :: Maybe ApplicationMetrics

Application request metrics for the environment.

httpStatus :: Int

The response's http status code.

newDescribeEnvironmentHealthResponse

Arguments

:: Int
$sel:httpStatus:DescribeEnvironmentHealthResponse'

-> DescribeEnvironmentHealthResponse

Create a value of DescribeEnvironmentHealthResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DescribeEnvironmentHealthResponse', describeEnvironmentHealthResponse_status - The environment's operational status. Ready, Launching, Updating, Terminating, or Terminated.

$sel:refreshedAt:DescribeEnvironmentHealthResponse', describeEnvironmentHealthResponse_refreshedAt - The date and time that the health information was retrieved.

$sel:color:DescribeEnvironmentHealthResponse', describeEnvironmentHealthResponse_color - The health color of the environment.

$sel:causes:DescribeEnvironmentHealthResponse', describeEnvironmentHealthResponse_causes - Descriptions of the data that contributed to the environment's current health status.

$sel:environmentName:DescribeEnvironmentHealth', describeEnvironmentHealthResponse_environmentName - The environment's name.

$sel:instancesHealth:DescribeEnvironmentHealthResponse', describeEnvironmentHealthResponse_instancesHealth - Summary health information for the instances in the environment.

$sel:healthStatus:DescribeEnvironmentHealthResponse', describeEnvironmentHealthResponse_healthStatus - The health status of the environment. For example, Ok.

$sel:applicationMetrics:DescribeEnvironmentHealthResponse', describeEnvironmentHealthResponse_applicationMetrics - Application request metrics for the environment.

$sel:httpStatus:DescribeEnvironmentHealthResponse', describeEnvironmentHealthResponse_httpStatus - The response's http status code.

Response Lenses

describeEnvironmentHealthResponse_status :: Lens' DescribeEnvironmentHealthResponse (Maybe EnvironmentHealth)

The environment's operational status. Ready, Launching, Updating, Terminating, or Terminated.

describeEnvironmentHealthResponse_refreshedAt :: Lens' DescribeEnvironmentHealthResponse (Maybe UTCTime)

The date and time that the health information was retrieved.

describeEnvironmentHealthResponse_color :: Lens' DescribeEnvironmentHealthResponse (Maybe Text)

The health color of the environment.

describeEnvironmentHealthResponse_causes :: Lens' DescribeEnvironmentHealthResponse (Maybe [Text])

Descriptions of the data that contributed to the environment's current health status.

describeEnvironmentHealthResponse_environmentName :: Lens' DescribeEnvironmentHealthResponse (Maybe Text)

The environment's name.

describeEnvironmentHealthResponse_instancesHealth :: Lens' DescribeEnvironmentHealthResponse (Maybe InstanceHealthSummary)

Summary health information for the instances in the environment.

describeEnvironmentHealthResponse_healthStatus :: Lens' DescribeEnvironmentHealthResponse (Maybe Text)

The health status of the environment. For example, Ok.

describeEnvironmentHealthResponse_applicationMetrics :: Lens' DescribeEnvironmentHealthResponse (Maybe ApplicationMetrics)

Application request metrics for the environment.

describeEnvironmentHealthResponse_httpStatus :: Lens' DescribeEnvironmentHealthResponse Int

The response's http status code.