Network.AWS.ElasticBeanstalk.DescribeEvents

Description

Returns list of event descriptions matching criteria up to the last 6 weeks.

This action returns the most recent 1,000 events from the specified NextToken.

This operation returns paginated results.

Creating a Request

data DescribeEvents

Request to retrieve a list of events for an environment.

See: newDescribeEvents smart constructor.

Constructors

DescribeEvents'

Fields

templateName :: Maybe Text

If specified, AWS Elastic Beanstalk restricts the returned descriptions to those that are associated with this environment configuration.

nextToken :: Maybe Text

Pagination token. If specified, the events return the next batch of results.

severity :: Maybe EventSeverity

If specified, limits the events returned from this call to include only those with the specified severity or higher.

environmentId :: Maybe Text

If specified, AWS Elastic Beanstalk restricts the returned descriptions to those associated with this environment.

startTime :: Maybe ISO8601

If specified, AWS Elastic Beanstalk restricts the returned descriptions to those that occur on or after this time.

platformArn :: Maybe Text

The ARN of a custom platform version. If specified, AWS Elastic Beanstalk restricts the returned descriptions to those associated with this custom platform version.

endTime :: Maybe ISO8601

If specified, AWS Elastic Beanstalk restricts the returned descriptions to those that occur up to, but not including, the EndTime.

environmentName :: Maybe Text

If specified, AWS Elastic Beanstalk restricts the returned descriptions to those associated with this environment.

versionLabel :: Maybe Text

If specified, AWS Elastic Beanstalk restricts the returned descriptions to those associated with this application version.

requestId :: Maybe Text

If specified, AWS Elastic Beanstalk restricts the described events to include only those associated with this request ID.

applicationName :: Maybe Text

If specified, AWS Elastic Beanstalk restricts the returned descriptions to include only those associated with this application.

maxRecords :: Maybe Natural

Specifies the maximum number of events that can be returned, beginning with the most recent event.

newDescribeEvents :: DescribeEvents

Create a value of DescribeEvents with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:templateName:DescribeEvents', describeEvents_templateName - If specified, AWS Elastic Beanstalk restricts the returned descriptions to those that are associated with this environment configuration.

$sel:nextToken:DescribeEvents', describeEvents_nextToken - Pagination token. If specified, the events return the next batch of results.

$sel:severity:DescribeEvents', describeEvents_severity - If specified, limits the events returned from this call to include only those with the specified severity or higher.

$sel:environmentId:DescribeEvents', describeEvents_environmentId - If specified, AWS Elastic Beanstalk restricts the returned descriptions to those associated with this environment.

$sel:startTime:DescribeEvents', describeEvents_startTime - If specified, AWS Elastic Beanstalk restricts the returned descriptions to those that occur on or after this time.

$sel:platformArn:DescribeEvents', describeEvents_platformArn - The ARN of a custom platform version. If specified, AWS Elastic Beanstalk restricts the returned descriptions to those associated with this custom platform version.

$sel:endTime:DescribeEvents', describeEvents_endTime - If specified, AWS Elastic Beanstalk restricts the returned descriptions to those that occur up to, but not including, the EndTime.

$sel:environmentName:DescribeEvents', describeEvents_environmentName - If specified, AWS Elastic Beanstalk restricts the returned descriptions to those associated with this environment.

$sel:versionLabel:DescribeEvents', describeEvents_versionLabel - If specified, AWS Elastic Beanstalk restricts the returned descriptions to those associated with this application version.

$sel:requestId:DescribeEvents', describeEvents_requestId - If specified, AWS Elastic Beanstalk restricts the described events to include only those associated with this request ID.

$sel:applicationName:DescribeEvents', describeEvents_applicationName - If specified, AWS Elastic Beanstalk restricts the returned descriptions to include only those associated with this application.

$sel:maxRecords:DescribeEvents', describeEvents_maxRecords - Specifies the maximum number of events that can be returned, beginning with the most recent event.

Request Lenses

describeEvents_templateName :: Lens' DescribeEvents (Maybe Text)

If specified, AWS Elastic Beanstalk restricts the returned descriptions to those that are associated with this environment configuration.

describeEvents_nextToken :: Lens' DescribeEvents (Maybe Text)

Pagination token. If specified, the events return the next batch of results.

describeEvents_severity :: Lens' DescribeEvents (Maybe EventSeverity)

If specified, limits the events returned from this call to include only those with the specified severity or higher.

describeEvents_environmentId :: Lens' DescribeEvents (Maybe Text)

If specified, AWS Elastic Beanstalk restricts the returned descriptions to those associated with this environment.

describeEvents_startTime :: Lens' DescribeEvents (Maybe UTCTime)

If specified, AWS Elastic Beanstalk restricts the returned descriptions to those that occur on or after this time.

describeEvents_platformArn :: Lens' DescribeEvents (Maybe Text)

The ARN of a custom platform version. If specified, AWS Elastic Beanstalk restricts the returned descriptions to those associated with this custom platform version.

describeEvents_endTime :: Lens' DescribeEvents (Maybe UTCTime)

If specified, AWS Elastic Beanstalk restricts the returned descriptions to those that occur up to, but not including, the EndTime.

describeEvents_environmentName :: Lens' DescribeEvents (Maybe Text)

If specified, AWS Elastic Beanstalk restricts the returned descriptions to those associated with this environment.

describeEvents_versionLabel :: Lens' DescribeEvents (Maybe Text)

If specified, AWS Elastic Beanstalk restricts the returned descriptions to those associated with this application version.

describeEvents_requestId :: Lens' DescribeEvents (Maybe Text)

If specified, AWS Elastic Beanstalk restricts the described events to include only those associated with this request ID.

describeEvents_applicationName :: Lens' DescribeEvents (Maybe Text)

If specified, AWS Elastic Beanstalk restricts the returned descriptions to include only those associated with this application.

describeEvents_maxRecords :: Lens' DescribeEvents (Maybe Natural)

Specifies the maximum number of events that can be returned, beginning with the most recent event.

Destructuring the Response

data DescribeEventsResponse

Result message wrapping a list of event descriptions.

See: newDescribeEventsResponse smart constructor.

Constructors

DescribeEventsResponse'

Fields

nextToken :: Maybe Text

If returned, this indicates that there are more results to obtain. Use this token in the next DescribeEvents call to get the next batch of events.

events :: Maybe [EventDescription]

A list of EventDescription.

httpStatus :: Int

The response's http status code.

newDescribeEventsResponse

Arguments

:: Int
$sel:httpStatus:DescribeEventsResponse'

-> DescribeEventsResponse

Create a value of DescribeEventsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeEvents', describeEventsResponse_nextToken - If returned, this indicates that there are more results to obtain. Use this token in the next DescribeEvents call to get the next batch of events.

$sel:events:DescribeEventsResponse', describeEventsResponse_events - A list of EventDescription.

$sel:httpStatus:DescribeEventsResponse', describeEventsResponse_httpStatus - The response's http status code.

Response Lenses

describeEventsResponse_nextToken :: Lens' DescribeEventsResponse (Maybe Text)

If returned, this indicates that there are more results to obtain. Use this token in the next DescribeEvents call to get the next batch of events.

describeEventsResponse_events :: Lens' DescribeEventsResponse (Maybe [EventDescription])

A list of EventDescription.

describeEventsResponse_httpStatus :: Lens' DescribeEventsResponse Int

The response's http status code.