Network.AWS.ElasticBeanstalk.DescribeInstancesHealth

Description

Retrieves detailed information about the health of instances in your AWS Elastic Beanstalk. This operation requires enhanced health reporting.

Creating a Request

data DescribeInstancesHealth

Parameters for a call to DescribeInstancesHealth.

See: newDescribeInstancesHealth smart constructor.

Constructors

DescribeInstancesHealth'

Fields

nextToken :: Maybe Text

Specify the pagination token returned by a previous call.

environmentId :: Maybe Text

Specify the AWS Elastic Beanstalk environment by ID.

environmentName :: Maybe Text

Specify the AWS Elastic Beanstalk environment by name.

attributeNames :: Maybe [InstancesHealthAttribute]

Specifies the response elements you wish to receive. To retrieve all attributes, set to All. If no attribute names are specified, returns a list of instances.

newDescribeInstancesHealth :: DescribeInstancesHealth

Create a value of DescribeInstancesHealth with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeInstancesHealth', describeInstancesHealth_nextToken - Specify the pagination token returned by a previous call.

$sel:environmentId:DescribeInstancesHealth', describeInstancesHealth_environmentId - Specify the AWS Elastic Beanstalk environment by ID.

$sel:environmentName:DescribeInstancesHealth', describeInstancesHealth_environmentName - Specify the AWS Elastic Beanstalk environment by name.

$sel:attributeNames:DescribeInstancesHealth', describeInstancesHealth_attributeNames - Specifies the response elements you wish to receive. To retrieve all attributes, set to All. If no attribute names are specified, returns a list of instances.

Request Lenses

describeInstancesHealth_nextToken :: Lens' DescribeInstancesHealth (Maybe Text)

Specify the pagination token returned by a previous call.

describeInstancesHealth_environmentId :: Lens' DescribeInstancesHealth (Maybe Text)

Specify the AWS Elastic Beanstalk environment by ID.

describeInstancesHealth_environmentName :: Lens' DescribeInstancesHealth (Maybe Text)

Specify the AWS Elastic Beanstalk environment by name.

describeInstancesHealth_attributeNames :: Lens' DescribeInstancesHealth (Maybe [InstancesHealthAttribute])

Specifies the response elements you wish to receive. To retrieve all attributes, set to All. If no attribute names are specified, returns a list of instances.

Destructuring the Response

data DescribeInstancesHealthResponse

Detailed health information about the Amazon EC2 instances in an AWS Elastic Beanstalk environment.

See: newDescribeInstancesHealthResponse smart constructor.

Constructors

DescribeInstancesHealthResponse'

Fields

instanceHealthList :: Maybe [SingleInstanceHealth]

Detailed health information about each instance.

The output differs slightly between Linux and Windows environments. There is a difference in the members that are supported under the <CPUUtilization> type.

nextToken :: Maybe Text

Pagination token for the next page of results, if available.

refreshedAt :: Maybe ISO8601

The date and time that the health information was retrieved.

httpStatus :: Int

The response's http status code.

newDescribeInstancesHealthResponse

Arguments

:: Int
$sel:httpStatus:DescribeInstancesHealthResponse'

-> DescribeInstancesHealthResponse

Create a value of DescribeInstancesHealthResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceHealthList:DescribeInstancesHealthResponse', describeInstancesHealthResponse_instanceHealthList - Detailed health information about each instance.

The output differs slightly between Linux and Windows environments. There is a difference in the members that are supported under the <CPUUtilization> type.

$sel:nextToken:DescribeInstancesHealth', describeInstancesHealthResponse_nextToken - Pagination token for the next page of results, if available.

$sel:refreshedAt:DescribeInstancesHealthResponse', describeInstancesHealthResponse_refreshedAt - The date and time that the health information was retrieved.

$sel:httpStatus:DescribeInstancesHealthResponse', describeInstancesHealthResponse_httpStatus - The response's http status code.

Response Lenses

describeInstancesHealthResponse_instanceHealthList :: Lens' DescribeInstancesHealthResponse (Maybe [SingleInstanceHealth])

Detailed health information about each instance.

The output differs slightly between Linux and Windows environments. There is a difference in the members that are supported under the <CPUUtilization> type.

describeInstancesHealthResponse_nextToken :: Lens' DescribeInstancesHealthResponse (Maybe Text)

Pagination token for the next page of results, if available.

describeInstancesHealthResponse_refreshedAt :: Lens' DescribeInstancesHealthResponse (Maybe UTCTime)

The date and time that the health information was retrieved.

describeInstancesHealthResponse_httpStatus :: Lens' DescribeInstancesHealthResponse Int

The response's http status code.