Network.AWS.ElasticBeanstalk.ListAvailableSolutionStacks

Description

Returns a list of the available solution stack names, with the public version first and then in reverse chronological order.

Creating a Request

data ListAvailableSolutionStacks

See: newListAvailableSolutionStacks smart constructor.

Constructors

ListAvailableSolutionStacks'

newListAvailableSolutionStacks :: ListAvailableSolutionStacks

Create a value of ListAvailableSolutionStacks with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

Destructuring the Response

data ListAvailableSolutionStacksResponse

A list of available AWS Elastic Beanstalk solution stacks.

See: newListAvailableSolutionStacksResponse smart constructor.

Constructors

ListAvailableSolutionStacksResponse'

Fields

solutionStacks :: Maybe [Text]

A list of available solution stacks.

solutionStackDetails :: Maybe [SolutionStackDescription]

A list of available solution stacks and their SolutionStackDescription.

httpStatus :: Int

The response's http status code.

newListAvailableSolutionStacksResponse

Arguments

:: Int
$sel:httpStatus:ListAvailableSolutionStacksResponse'

-> ListAvailableSolutionStacksResponse

Create a value of ListAvailableSolutionStacksResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:solutionStacks:ListAvailableSolutionStacksResponse', listAvailableSolutionStacksResponse_solutionStacks - A list of available solution stacks.

$sel:solutionStackDetails:ListAvailableSolutionStacksResponse', listAvailableSolutionStacksResponse_solutionStackDetails - A list of available solution stacks and their SolutionStackDescription.

$sel:httpStatus:ListAvailableSolutionStacksResponse', listAvailableSolutionStacksResponse_httpStatus - The response's http status code.

Response Lenses

listAvailableSolutionStacksResponse_solutionStacks :: Lens' ListAvailableSolutionStacksResponse (Maybe [Text])

A list of available solution stacks.

listAvailableSolutionStacksResponse_solutionStackDetails :: Lens' ListAvailableSolutionStacksResponse (Maybe [SolutionStackDescription])

A list of available solution stacks and their SolutionStackDescription.

listAvailableSolutionStacksResponse_httpStatus :: Lens' ListAvailableSolutionStacksResponse Int

The response's http status code.