Network.AWS.ElasticBeanstalk.ListPlatformVersions

Description

Lists the platform versions available for your account in an AWS Region. Provides summary information about each platform version. Compare to DescribePlatformVersion, which provides full details about a single platform version.

For definitions of platform version and other platform-related terms, see AWS Elastic Beanstalk Platforms Glossary.

This operation returns paginated results.

Creating a Request

data ListPlatformVersions

See: newListPlatformVersions smart constructor.

Constructors

ListPlatformVersions'

Fields

filters :: Maybe [PlatformFilter]

Criteria for restricting the resulting list of platform versions. The filter is interpreted as a logical conjunction (AND) of the separate PlatformFilter terms.

nextToken :: Maybe Text

For a paginated request. Specify a token from a previous response page to retrieve the next response page. All other parameter values must be identical to the ones specified in the initial request.

If no NextToken is specified, the first page is retrieved.

maxRecords :: Maybe Natural

The maximum number of platform version values returned in one call.

newListPlatformVersions :: ListPlatformVersions

Create a value of ListPlatformVersions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:ListPlatformVersions', listPlatformVersions_filters - Criteria for restricting the resulting list of platform versions. The filter is interpreted as a logical conjunction (AND) of the separate PlatformFilter terms.

$sel:nextToken:ListPlatformVersions', listPlatformVersions_nextToken - For a paginated request. Specify a token from a previous response page to retrieve the next response page. All other parameter values must be identical to the ones specified in the initial request.

If no NextToken is specified, the first page is retrieved.

$sel:maxRecords:ListPlatformVersions', listPlatformVersions_maxRecords - The maximum number of platform version values returned in one call.

Request Lenses

listPlatformVersions_filters :: Lens' ListPlatformVersions (Maybe [PlatformFilter])

Criteria for restricting the resulting list of platform versions. The filter is interpreted as a logical conjunction (AND) of the separate PlatformFilter terms.

listPlatformVersions_nextToken :: Lens' ListPlatformVersions (Maybe Text)

For a paginated request. Specify a token from a previous response page to retrieve the next response page. All other parameter values must be identical to the ones specified in the initial request.

If no NextToken is specified, the first page is retrieved.

listPlatformVersions_maxRecords :: Lens' ListPlatformVersions (Maybe Natural)

The maximum number of platform version values returned in one call.

Destructuring the Response

data ListPlatformVersionsResponse

See: newListPlatformVersionsResponse smart constructor.

Constructors

ListPlatformVersionsResponse'

Fields

nextToken :: Maybe Text

In a paginated request, if this value isn't null, it's the token that you can pass in a subsequent request to get the next response page.

platformSummaryList :: Maybe [PlatformSummary]

Summary information about the platform versions.

httpStatus :: Int

The response's http status code.

newListPlatformVersionsResponse

Arguments

:: Int
$sel:httpStatus:ListPlatformVersionsResponse'

-> ListPlatformVersionsResponse

Create a value of ListPlatformVersionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListPlatformVersions', listPlatformVersionsResponse_nextToken - In a paginated request, if this value isn't null, it's the token that you can pass in a subsequent request to get the next response page.

$sel:platformSummaryList:ListPlatformVersionsResponse', listPlatformVersionsResponse_platformSummaryList - Summary information about the platform versions.

$sel:httpStatus:ListPlatformVersionsResponse', listPlatformVersionsResponse_httpStatus - The response's http status code.

Response Lenses

listPlatformVersionsResponse_nextToken :: Lens' ListPlatformVersionsResponse (Maybe Text)

In a paginated request, if this value isn't null, it's the token that you can pass in a subsequent request to get the next response page.

listPlatformVersionsResponse_platformSummaryList :: Lens' ListPlatformVersionsResponse (Maybe [PlatformSummary])

Summary information about the platform versions.

listPlatformVersionsResponse_httpStatus :: Lens' ListPlatformVersionsResponse Int

The response's http status code.