Network.AWS.ElasticBeanstalk.RestartAppServer

Description

Causes the environment to restart the application container server running on each Amazon EC2 instance.

Creating a Request

data RestartAppServer

See: newRestartAppServer smart constructor.

Constructors

RestartAppServer'

Fields

environmentId :: Maybe Text

The ID of the environment to restart the server for.

Condition: You must specify either this or an EnvironmentName, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

environmentName :: Maybe Text

The name of the environment to restart the server for.

Condition: You must specify either this or an EnvironmentId, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

newRestartAppServer :: RestartAppServer

Create a value of RestartAppServer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:environmentId:RestartAppServer', restartAppServer_environmentId - The ID of the environment to restart the server for.

Condition: You must specify either this or an EnvironmentName, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

$sel:environmentName:RestartAppServer', restartAppServer_environmentName - The name of the environment to restart the server for.

Condition: You must specify either this or an EnvironmentId, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

Request Lenses

restartAppServer_environmentId :: Lens' RestartAppServer (Maybe Text)

The ID of the environment to restart the server for.

Condition: You must specify either this or an EnvironmentName, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

restartAppServer_environmentName :: Lens' RestartAppServer (Maybe Text)

The name of the environment to restart the server for.

Condition: You must specify either this or an EnvironmentId, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

Destructuring the Response

data RestartAppServerResponse

See: newRestartAppServerResponse smart constructor.

Constructors

RestartAppServerResponse'

newRestartAppServerResponse :: RestartAppServerResponse

Create a value of RestartAppServerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.