Network.AWS.ElasticBeanstalk.Types.ApplicationDescription

Description

Documentation

data ApplicationDescription

Describes the properties of an application.

See: newApplicationDescription smart constructor.

Constructors

ApplicationDescription'

Fields

applicationArn :: Maybe Text

The Amazon Resource Name (ARN) of the application.

dateCreated :: Maybe ISO8601

The date when the application was created.

versions :: Maybe [Text]

The names of the versions for this application.

dateUpdated :: Maybe ISO8601

The date when the application was last modified.

resourceLifecycleConfig :: Maybe ApplicationResourceLifecycleConfig

The lifecycle settings for the application.

description :: Maybe Text

User-defined description of the application.

configurationTemplates :: Maybe [Text]

The names of the configuration templates associated with this application.

applicationName :: Maybe Text

The name of the application.

newApplicationDescription :: ApplicationDescription

Create a value of ApplicationDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationArn:ApplicationDescription', applicationDescription_applicationArn - The Amazon Resource Name (ARN) of the application.

$sel:dateCreated:ApplicationDescription', applicationDescription_dateCreated - The date when the application was created.

$sel:versions:ApplicationDescription', applicationDescription_versions - The names of the versions for this application.

$sel:dateUpdated:ApplicationDescription', applicationDescription_dateUpdated - The date when the application was last modified.

$sel:resourceLifecycleConfig:ApplicationDescription', applicationDescription_resourceLifecycleConfig - The lifecycle settings for the application.

$sel:description:ApplicationDescription', applicationDescription_description - User-defined description of the application.

$sel:configurationTemplates:ApplicationDescription', applicationDescription_configurationTemplates - The names of the configuration templates associated with this application.

$sel:applicationName:ApplicationDescription', applicationDescription_applicationName - The name of the application.

applicationDescription_applicationArn :: Lens' ApplicationDescription (Maybe Text)

The Amazon Resource Name (ARN) of the application.

applicationDescription_dateCreated :: Lens' ApplicationDescription (Maybe UTCTime)

The date when the application was created.

applicationDescription_versions :: Lens' ApplicationDescription (Maybe [Text])

The names of the versions for this application.

applicationDescription_dateUpdated :: Lens' ApplicationDescription (Maybe UTCTime)

The date when the application was last modified.

applicationDescription_resourceLifecycleConfig :: Lens' ApplicationDescription (Maybe ApplicationResourceLifecycleConfig)

The lifecycle settings for the application.

applicationDescription_description :: Lens' ApplicationDescription (Maybe Text)

User-defined description of the application.

applicationDescription_configurationTemplates :: Lens' ApplicationDescription (Maybe [Text])

The names of the configuration templates associated with this application.

applicationDescription_applicationName :: Lens' ApplicationDescription (Maybe Text)

The name of the application.