Network.AWS.ElasticBeanstalk.Types.ApplicationVersionLifecycleConfig

Description

Documentation

data ApplicationVersionLifecycleConfig

The application version lifecycle settings for an application. Defines the rules that Elastic Beanstalk applies to an application's versions in order to avoid hitting the per-region limit for application versions.

When Elastic Beanstalk deletes an application version from its database, you can no longer deploy that version to an environment. The source bundle remains in S3 unless you configure the rule to delete it.

See: newApplicationVersionLifecycleConfig smart constructor.

Constructors

ApplicationVersionLifecycleConfig'

Fields

maxAgeRule :: Maybe MaxAgeRule

Specify a max age rule to restrict the length of time that application versions are retained for an application.

maxCountRule :: Maybe MaxCountRule

Specify a max count rule to restrict the number of application versions that are retained for an application.

newApplicationVersionLifecycleConfig :: ApplicationVersionLifecycleConfig

Create a value of ApplicationVersionLifecycleConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxAgeRule:ApplicationVersionLifecycleConfig', applicationVersionLifecycleConfig_maxAgeRule - Specify a max age rule to restrict the length of time that application versions are retained for an application.

$sel:maxCountRule:ApplicationVersionLifecycleConfig', applicationVersionLifecycleConfig_maxCountRule - Specify a max count rule to restrict the number of application versions that are retained for an application.

applicationVersionLifecycleConfig_maxAgeRule :: Lens' ApplicationVersionLifecycleConfig (Maybe MaxAgeRule)

Specify a max age rule to restrict the length of time that application versions are retained for an application.

applicationVersionLifecycleConfig_maxCountRule :: Lens' ApplicationVersionLifecycleConfig (Maybe MaxCountRule)

Specify a max count rule to restrict the number of application versions that are retained for an application.