Network.AWS.ElasticBeanstalk.Types.BuildConfiguration

Description

Documentation

data BuildConfiguration

Settings for an AWS CodeBuild build.

See: newBuildConfiguration smart constructor.

Constructors

BuildConfiguration'

Fields

artifactName :: Maybe Text

The name of the artifact of the CodeBuild build. If provided, Elastic Beanstalk stores the build artifact in the S3 location S3-bucket/resources/application-name/codebuild/codebuild-version-label-artifact-name.zip. If not provided, Elastic Beanstalk stores the build artifact in the S3 location S3-bucket/resources/application-name/codebuild/codebuild-version-label.zip.

timeoutInMinutes :: Maybe Int

How long in minutes, from 5 to 480 (8 hours), for AWS CodeBuild to wait until timing out any related build that does not get marked as completed. The default is 60 minutes.

computeType :: Maybe ComputeType

Information about the compute resources the build project will use.

BUILD_GENERAL1_SMALL: Use up to 3 GB memory and 2 vCPUs for builds

BUILD_GENERAL1_MEDIUM: Use up to 7 GB memory and 4 vCPUs for builds

BUILD_GENERAL1_LARGE: Use up to 15 GB memory and 8 vCPUs for builds

codeBuildServiceRole :: Text

The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role that enables AWS CodeBuild to interact with dependent AWS services on behalf of the AWS account.

image :: Text

The ID of the Docker image to use for this build project.

newBuildConfiguration

Arguments

:: Text
$sel:codeBuildServiceRole:BuildConfiguration'

-> Text
$sel:image:BuildConfiguration'

-> BuildConfiguration

Create a value of BuildConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:artifactName:BuildConfiguration', buildConfiguration_artifactName - The name of the artifact of the CodeBuild build. If provided, Elastic Beanstalk stores the build artifact in the S3 location S3-bucket/resources/application-name/codebuild/codebuild-version-label-artifact-name.zip. If not provided, Elastic Beanstalk stores the build artifact in the S3 location S3-bucket/resources/application-name/codebuild/codebuild-version-label.zip.

$sel:timeoutInMinutes:BuildConfiguration', buildConfiguration_timeoutInMinutes - How long in minutes, from 5 to 480 (8 hours), for AWS CodeBuild to wait until timing out any related build that does not get marked as completed. The default is 60 minutes.

$sel:computeType:BuildConfiguration', buildConfiguration_computeType - Information about the compute resources the build project will use.

BUILD_GENERAL1_SMALL: Use up to 3 GB memory and 2 vCPUs for builds

BUILD_GENERAL1_MEDIUM: Use up to 7 GB memory and 4 vCPUs for builds

BUILD_GENERAL1_LARGE: Use up to 15 GB memory and 8 vCPUs for builds

$sel:codeBuildServiceRole:BuildConfiguration', buildConfiguration_codeBuildServiceRole - The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role that enables AWS CodeBuild to interact with dependent AWS services on behalf of the AWS account.

$sel:image:BuildConfiguration', buildConfiguration_image - The ID of the Docker image to use for this build project.

buildConfiguration_artifactName :: Lens' BuildConfiguration (Maybe Text)

The name of the artifact of the CodeBuild build. If provided, Elastic Beanstalk stores the build artifact in the S3 location S3-bucket/resources/application-name/codebuild/codebuild-version-label-artifact-name.zip. If not provided, Elastic Beanstalk stores the build artifact in the S3 location S3-bucket/resources/application-name/codebuild/codebuild-version-label.zip.

buildConfiguration_timeoutInMinutes :: Lens' BuildConfiguration (Maybe Int)

How long in minutes, from 5 to 480 (8 hours), for AWS CodeBuild to wait until timing out any related build that does not get marked as completed. The default is 60 minutes.

buildConfiguration_computeType :: Lens' BuildConfiguration (Maybe ComputeType)

Information about the compute resources the build project will use.

BUILD_GENERAL1_SMALL: Use up to 3 GB memory and 2 vCPUs for builds

BUILD_GENERAL1_MEDIUM: Use up to 7 GB memory and 4 vCPUs for builds

BUILD_GENERAL1_LARGE: Use up to 15 GB memory and 8 vCPUs for builds

buildConfiguration_codeBuildServiceRole :: Lens' BuildConfiguration Text

The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role that enables AWS CodeBuild to interact with dependent AWS services on behalf of the AWS account.

buildConfiguration_image :: Lens' BuildConfiguration Text

The ID of the Docker image to use for this build project.