Network.AWS.ElasticBeanstalk.Types.CPUUtilization

Description

Documentation

data CPUUtilization

CPU utilization metrics for an instance.

See: newCPUUtilization smart constructor.

Constructors

CPUUtilization'

Fields

softIRQ :: Maybe Double

Available on Linux environments only.

Percentage of time that the CPU has spent in the SoftIRQ state over the last 10 seconds.

idle :: Maybe Double

Percentage of time that the CPU has spent in the Idle state over the last 10 seconds.

irq :: Maybe Double

Available on Linux environments only.

Percentage of time that the CPU has spent in the IRQ state over the last 10 seconds.

system :: Maybe Double

Available on Linux environments only.

Percentage of time that the CPU has spent in the System state over the last 10 seconds.

privileged :: Maybe Double

Available on Windows environments only.

Percentage of time that the CPU has spent in the Privileged state over the last 10 seconds.

user :: Maybe Double

Percentage of time that the CPU has spent in the User state over the last 10 seconds.

iOWait :: Maybe Double

Available on Linux environments only.

Percentage of time that the CPU has spent in the I/O Wait state over the last 10 seconds.

nice :: Maybe Double

Available on Linux environments only.

Percentage of time that the CPU has spent in the Nice state over the last 10 seconds.

newCPUUtilization :: CPUUtilization

Create a value of CPUUtilization with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:softIRQ:CPUUtilization', cPUUtilization_softIRQ - Available on Linux environments only.

Percentage of time that the CPU has spent in the SoftIRQ state over the last 10 seconds.

$sel:idle:CPUUtilization', cPUUtilization_idle - Percentage of time that the CPU has spent in the Idle state over the last 10 seconds.

$sel:irq:CPUUtilization', cPUUtilization_irq - Available on Linux environments only.

Percentage of time that the CPU has spent in the IRQ state over the last 10 seconds.

$sel:system:CPUUtilization', cPUUtilization_system - Available on Linux environments only.

Percentage of time that the CPU has spent in the System state over the last 10 seconds.

$sel:privileged:CPUUtilization', cPUUtilization_privileged - Available on Windows environments only.

Percentage of time that the CPU has spent in the Privileged state over the last 10 seconds.

$sel:user:CPUUtilization', cPUUtilization_user - Percentage of time that the CPU has spent in the User state over the last 10 seconds.

$sel:iOWait:CPUUtilization', cPUUtilization_iOWait - Available on Linux environments only.

Percentage of time that the CPU has spent in the I/O Wait state over the last 10 seconds.

$sel:nice:CPUUtilization', cPUUtilization_nice - Available on Linux environments only.

Percentage of time that the CPU has spent in the Nice state over the last 10 seconds.

cPUUtilization_softIRQ :: Lens' CPUUtilization (Maybe Double)

Available on Linux environments only.

Percentage of time that the CPU has spent in the SoftIRQ state over the last 10 seconds.

cPUUtilization_idle :: Lens' CPUUtilization (Maybe Double)

Percentage of time that the CPU has spent in the Idle state over the last 10 seconds.

cPUUtilization_irq :: Lens' CPUUtilization (Maybe Double)

Available on Linux environments only.

Percentage of time that the CPU has spent in the IRQ state over the last 10 seconds.

cPUUtilization_system :: Lens' CPUUtilization (Maybe Double)

Available on Linux environments only.

Percentage of time that the CPU has spent in the System state over the last 10 seconds.

cPUUtilization_privileged :: Lens' CPUUtilization (Maybe Double)

Available on Windows environments only.

Percentage of time that the CPU has spent in the Privileged state over the last 10 seconds.

cPUUtilization_user :: Lens' CPUUtilization (Maybe Double)

Percentage of time that the CPU has spent in the User state over the last 10 seconds.

cPUUtilization_iOWait :: Lens' CPUUtilization (Maybe Double)

Available on Linux environments only.

Percentage of time that the CPU has spent in the I/O Wait state over the last 10 seconds.

cPUUtilization_nice :: Lens' CPUUtilization (Maybe Double)

Available on Linux environments only.

Percentage of time that the CPU has spent in the Nice state over the last 10 seconds.