Network.AWS.ElasticBeanstalk.Types.ConfigurationOptionDescription

Description

Documentation

data ConfigurationOptionDescription

Describes the possible values for a configuration option.

See: newConfigurationOptionDescription smart constructor.

Constructors

ConfigurationOptionDescription'

Fields

maxValue :: Maybe Int

If specified, the configuration option must be a numeric value less than this value.

valueOptions :: Maybe [Text]

If specified, values for the configuration option are selected from this list.

valueType :: Maybe ConfigurationOptionValueType

An indication of which type of values this option has and whether it is allowable to select one or more than one of the possible values:

Scalar : Values for this option are a single selection from the possible values, or an unformatted string, or numeric value governed by the MIN/MAX/Regex constraints.

List : Values for this option are multiple selections from the possible values.

Boolean : Values for this option are either true or false .

Json : Values for this option are a JSON representation of a ConfigDocument.

changeSeverity :: Maybe Text

An indication of which action is required if the value for this configuration option changes:

NoInterruption : There is no interruption to the environment or application availability.

RestartEnvironment : The environment is entirely restarted, all AWS resources are deleted and recreated, and the environment is unavailable during the process.

RestartApplicationServer : The environment is available the entire time. However, a short application outage occurs when the application servers on the running Amazon EC2 instances are restarted.

regex :: Maybe OptionRestrictionRegex

If specified, the configuration option must be a string value that satisfies this regular expression.

name :: Maybe Text

The name of the configuration option.

minValue :: Maybe Int

If specified, the configuration option must be a numeric value greater than this value.

namespace :: Maybe Text

A unique namespace identifying the option's associated AWS resource.

maxLength :: Maybe Int

If specified, the configuration option must be a string value no longer than this value.

defaultValue :: Maybe Text

The default value for this configuration option.

userDefined :: Maybe Bool

An indication of whether the user defined this configuration option:

true : This configuration option was defined by the user. It is a valid choice for specifying if this as an Option to Remove when updating configuration settings.

false : This configuration was not defined by the user.

Constraint: You can remove only UserDefined options from a configuration.

Valid Values: true | false

newConfigurationOptionDescription :: ConfigurationOptionDescription

Create a value of ConfigurationOptionDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxValue:ConfigurationOptionDescription', configurationOptionDescription_maxValue - If specified, the configuration option must be a numeric value less than this value.

$sel:valueOptions:ConfigurationOptionDescription', configurationOptionDescription_valueOptions - If specified, values for the configuration option are selected from this list.

$sel:valueType:ConfigurationOptionDescription', configurationOptionDescription_valueType - An indication of which type of values this option has and whether it is allowable to select one or more than one of the possible values:

Scalar : Values for this option are a single selection from the possible values, or an unformatted string, or numeric value governed by the MIN/MAX/Regex constraints.

List : Values for this option are multiple selections from the possible values.

Boolean : Values for this option are either true or false .

Json : Values for this option are a JSON representation of a ConfigDocument.

$sel:changeSeverity:ConfigurationOptionDescription', configurationOptionDescription_changeSeverity - An indication of which action is required if the value for this configuration option changes:

NoInterruption : There is no interruption to the environment or application availability.

RestartEnvironment : The environment is entirely restarted, all AWS resources are deleted and recreated, and the environment is unavailable during the process.

RestartApplicationServer : The environment is available the entire time. However, a short application outage occurs when the application servers on the running Amazon EC2 instances are restarted.

$sel:regex:ConfigurationOptionDescription', configurationOptionDescription_regex - If specified, the configuration option must be a string value that satisfies this regular expression.

$sel:name:ConfigurationOptionDescription', configurationOptionDescription_name - The name of the configuration option.

$sel:minValue:ConfigurationOptionDescription', configurationOptionDescription_minValue - If specified, the configuration option must be a numeric value greater than this value.

$sel:namespace:ConfigurationOptionDescription', configurationOptionDescription_namespace - A unique namespace identifying the option's associated AWS resource.

$sel:maxLength:ConfigurationOptionDescription', configurationOptionDescription_maxLength - If specified, the configuration option must be a string value no longer than this value.

$sel:defaultValue:ConfigurationOptionDescription', configurationOptionDescription_defaultValue - The default value for this configuration option.

$sel:userDefined:ConfigurationOptionDescription', configurationOptionDescription_userDefined - An indication of whether the user defined this configuration option:

true : This configuration option was defined by the user. It is a valid choice for specifying if this as an Option to Remove when updating configuration settings.

false : This configuration was not defined by the user.

Constraint: You can remove only UserDefined options from a configuration.

Valid Values: true | false

configurationOptionDescription_maxValue :: Lens' ConfigurationOptionDescription (Maybe Int)

If specified, the configuration option must be a numeric value less than this value.

configurationOptionDescription_valueOptions :: Lens' ConfigurationOptionDescription (Maybe [Text])

If specified, values for the configuration option are selected from this list.

configurationOptionDescription_valueType :: Lens' ConfigurationOptionDescription (Maybe ConfigurationOptionValueType)

An indication of which type of values this option has and whether it is allowable to select one or more than one of the possible values:

Scalar : Values for this option are a single selection from the possible values, or an unformatted string, or numeric value governed by the MIN/MAX/Regex constraints.

List : Values for this option are multiple selections from the possible values.

Boolean : Values for this option are either true or false .

Json : Values for this option are a JSON representation of a ConfigDocument.

configurationOptionDescription_changeSeverity :: Lens' ConfigurationOptionDescription (Maybe Text)

An indication of which action is required if the value for this configuration option changes:

NoInterruption : There is no interruption to the environment or application availability.

RestartEnvironment : The environment is entirely restarted, all AWS resources are deleted and recreated, and the environment is unavailable during the process.

RestartApplicationServer : The environment is available the entire time. However, a short application outage occurs when the application servers on the running Amazon EC2 instances are restarted.

configurationOptionDescription_regex :: Lens' ConfigurationOptionDescription (Maybe OptionRestrictionRegex)

If specified, the configuration option must be a string value that satisfies this regular expression.

configurationOptionDescription_name :: Lens' ConfigurationOptionDescription (Maybe Text)

The name of the configuration option.

configurationOptionDescription_minValue :: Lens' ConfigurationOptionDescription (Maybe Int)

If specified, the configuration option must be a numeric value greater than this value.

configurationOptionDescription_namespace :: Lens' ConfigurationOptionDescription (Maybe Text)

A unique namespace identifying the option's associated AWS resource.

configurationOptionDescription_maxLength :: Lens' ConfigurationOptionDescription (Maybe Int)

If specified, the configuration option must be a string value no longer than this value.

configurationOptionDescription_defaultValue :: Lens' ConfigurationOptionDescription (Maybe Text)

The default value for this configuration option.

configurationOptionDescription_userDefined :: Lens' ConfigurationOptionDescription (Maybe Bool)

An indication of whether the user defined this configuration option:

true : This configuration option was defined by the user. It is a valid choice for specifying if this as an Option to Remove when updating configuration settings.

false : This configuration was not defined by the user.

Constraint: You can remove only UserDefined options from a configuration.

Valid Values: true | false