Network.AWS.ElasticBeanstalk.Types.ConfigurationSettingsDescription

Description

Documentation

data ConfigurationSettingsDescription

Describes the settings for a configuration set.

See: newConfigurationSettingsDescription smart constructor.

Constructors

ConfigurationSettingsDescription'

Fields

templateName :: Maybe Text

If not null, the name of the configuration template for this configuration set.

dateCreated :: Maybe ISO8601

The date (in UTC time) when this configuration set was created.

solutionStackName :: Maybe Text

The name of the solution stack this configuration set uses.

deploymentStatus :: Maybe ConfigurationDeploymentStatus

If this configuration set is associated with an environment, the DeploymentStatus parameter indicates the deployment status of this configuration set:

null: This configuration is not associated with a running environment.

pending: This is a draft configuration that is not deployed to the associated environment but is in the process of deploying.

deployed: This is the configuration that is currently deployed to the associated running environment.

failed: This is a draft configuration that failed to successfully deploy.

platformArn :: Maybe Text

The ARN of the platform version.

environmentName :: Maybe Text

If not null, the name of the environment for this configuration set.

dateUpdated :: Maybe ISO8601

The date (in UTC time) when this configuration set was last modified.

optionSettings :: Maybe [ConfigurationOptionSetting]

A list of the configuration options and their values in this configuration set.

description :: Maybe Text

Describes this configuration set.

applicationName :: Maybe Text

The name of the application associated with this configuration set.

newConfigurationSettingsDescription :: ConfigurationSettingsDescription

Create a value of ConfigurationSettingsDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:templateName:ConfigurationSettingsDescription', configurationSettingsDescription_templateName - If not null, the name of the configuration template for this configuration set.

$sel:dateCreated:ConfigurationSettingsDescription', configurationSettingsDescription_dateCreated - The date (in UTC time) when this configuration set was created.

$sel:solutionStackName:ConfigurationSettingsDescription', configurationSettingsDescription_solutionStackName - The name of the solution stack this configuration set uses.

$sel:deploymentStatus:ConfigurationSettingsDescription', configurationSettingsDescription_deploymentStatus - If this configuration set is associated with an environment, the DeploymentStatus parameter indicates the deployment status of this configuration set:

null: This configuration is not associated with a running environment.

pending: This is a draft configuration that is not deployed to the associated environment but is in the process of deploying.

deployed: This is the configuration that is currently deployed to the associated running environment.

failed: This is a draft configuration that failed to successfully deploy.

$sel:platformArn:ConfigurationSettingsDescription', configurationSettingsDescription_platformArn - The ARN of the platform version.

$sel:environmentName:ConfigurationSettingsDescription', configurationSettingsDescription_environmentName - If not null, the name of the environment for this configuration set.

$sel:dateUpdated:ConfigurationSettingsDescription', configurationSettingsDescription_dateUpdated - The date (in UTC time) when this configuration set was last modified.

$sel:optionSettings:ConfigurationSettingsDescription', configurationSettingsDescription_optionSettings - A list of the configuration options and their values in this configuration set.

$sel:description:ConfigurationSettingsDescription', configurationSettingsDescription_description - Describes this configuration set.

$sel:applicationName:ConfigurationSettingsDescription', configurationSettingsDescription_applicationName - The name of the application associated with this configuration set.

configurationSettingsDescription_templateName :: Lens' ConfigurationSettingsDescription (Maybe Text)

If not null, the name of the configuration template for this configuration set.

configurationSettingsDescription_dateCreated :: Lens' ConfigurationSettingsDescription (Maybe UTCTime)

The date (in UTC time) when this configuration set was created.

configurationSettingsDescription_solutionStackName :: Lens' ConfigurationSettingsDescription (Maybe Text)

The name of the solution stack this configuration set uses.

configurationSettingsDescription_deploymentStatus :: Lens' ConfigurationSettingsDescription (Maybe ConfigurationDeploymentStatus)

If this configuration set is associated with an environment, the DeploymentStatus parameter indicates the deployment status of this configuration set:

null: This configuration is not associated with a running environment.

pending: This is a draft configuration that is not deployed to the associated environment but is in the process of deploying.

deployed: This is the configuration that is currently deployed to the associated running environment.

failed: This is a draft configuration that failed to successfully deploy.

configurationSettingsDescription_platformArn :: Lens' ConfigurationSettingsDescription (Maybe Text)

The ARN of the platform version.

configurationSettingsDescription_environmentName :: Lens' ConfigurationSettingsDescription (Maybe Text)

If not null, the name of the environment for this configuration set.

configurationSettingsDescription_dateUpdated :: Lens' ConfigurationSettingsDescription (Maybe UTCTime)

The date (in UTC time) when this configuration set was last modified.

configurationSettingsDescription_optionSettings :: Lens' ConfigurationSettingsDescription (Maybe [ConfigurationOptionSetting])

A list of the configuration options and their values in this configuration set.

configurationSettingsDescription_description :: Lens' ConfigurationSettingsDescription (Maybe Text)

Describes this configuration set.

configurationSettingsDescription_applicationName :: Lens' ConfigurationSettingsDescription (Maybe Text)

The name of the application associated with this configuration set.