Network.AWS.ElasticBeanstalk.Types.EnvironmentInfoDescription

Description

Documentation

data EnvironmentInfoDescription

The information retrieved from the Amazon EC2 instances.

See: newEnvironmentInfoDescription smart constructor.

Constructors

EnvironmentInfoDescription'

Fields

message :: Maybe Text

The retrieved information. Currently contains a presigned Amazon S3 URL. The files are deleted after 15 minutes.

Anyone in possession of this URL can access the files before they are deleted. Make the URL available only to trusted parties.

infoType :: Maybe EnvironmentInfoType

The type of information retrieved.

sampleTimestamp :: Maybe ISO8601

The time stamp when this information was retrieved.

ec2InstanceId :: Maybe Text

The Amazon EC2 Instance ID for this information.

newEnvironmentInfoDescription :: EnvironmentInfoDescription

Create a value of EnvironmentInfoDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:message:EnvironmentInfoDescription', environmentInfoDescription_message - The retrieved information. Currently contains a presigned Amazon S3 URL. The files are deleted after 15 minutes.

Anyone in possession of this URL can access the files before they are deleted. Make the URL available only to trusted parties.

$sel:infoType:EnvironmentInfoDescription', environmentInfoDescription_infoType - The type of information retrieved.

$sel:sampleTimestamp:EnvironmentInfoDescription', environmentInfoDescription_sampleTimestamp - The time stamp when this information was retrieved.

$sel:ec2InstanceId:EnvironmentInfoDescription', environmentInfoDescription_ec2InstanceId - The Amazon EC2 Instance ID for this information.

environmentInfoDescription_message :: Lens' EnvironmentInfoDescription (Maybe Text)

The retrieved information. Currently contains a presigned Amazon S3 URL. The files are deleted after 15 minutes.

Anyone in possession of this URL can access the files before they are deleted. Make the URL available only to trusted parties.

environmentInfoDescription_infoType :: Lens' EnvironmentInfoDescription (Maybe EnvironmentInfoType)

The type of information retrieved.

environmentInfoDescription_sampleTimestamp :: Lens' EnvironmentInfoDescription (Maybe UTCTime)

The time stamp when this information was retrieved.

environmentInfoDescription_ec2InstanceId :: Lens' EnvironmentInfoDescription (Maybe Text)

The Amazon EC2 Instance ID for this information.