Network.AWS.ElasticBeanstalk.Types.EnvironmentResourceDescription

Description

Documentation

data EnvironmentResourceDescription

Describes the AWS resources in use by this environment. This data is live.

See: newEnvironmentResourceDescription smart constructor.

Constructors

EnvironmentResourceDescription'

Fields

launchTemplates :: Maybe [LaunchTemplate]

The Amazon EC2 launch templates in use by this environment.

launchConfigurations :: Maybe [LaunchConfiguration]

The Auto Scaling launch configurations in use by this environment.

triggers :: Maybe [Trigger]

The AutoScaling triggers in use by this environment.

queues :: Maybe [Queue]

The queues used by this environment.

instances :: Maybe [Instance]

The Amazon EC2 instances used by this environment.

environmentName :: Maybe Text

The name of the environment.

loadBalancers :: Maybe [LoadBalancer]

The LoadBalancers in use by this environment.

autoScalingGroups :: Maybe [AutoScalingGroup]

The AutoScalingGroups used by this environment.

newEnvironmentResourceDescription :: EnvironmentResourceDescription

Create a value of EnvironmentResourceDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:launchTemplates:EnvironmentResourceDescription', environmentResourceDescription_launchTemplates - The Amazon EC2 launch templates in use by this environment.

$sel:launchConfigurations:EnvironmentResourceDescription', environmentResourceDescription_launchConfigurations - The Auto Scaling launch configurations in use by this environment.

$sel:triggers:EnvironmentResourceDescription', environmentResourceDescription_triggers - The AutoScaling triggers in use by this environment.

$sel:queues:EnvironmentResourceDescription', environmentResourceDescription_queues - The queues used by this environment.

$sel:instances:EnvironmentResourceDescription', environmentResourceDescription_instances - The Amazon EC2 instances used by this environment.

$sel:environmentName:EnvironmentResourceDescription', environmentResourceDescription_environmentName - The name of the environment.

$sel:loadBalancers:EnvironmentResourceDescription', environmentResourceDescription_loadBalancers - The LoadBalancers in use by this environment.

$sel:autoScalingGroups:EnvironmentResourceDescription', environmentResourceDescription_autoScalingGroups - The AutoScalingGroups used by this environment.

environmentResourceDescription_launchTemplates :: Lens' EnvironmentResourceDescription (Maybe [LaunchTemplate])

The Amazon EC2 launch templates in use by this environment.

environmentResourceDescription_launchConfigurations :: Lens' EnvironmentResourceDescription (Maybe [LaunchConfiguration])

The Auto Scaling launch configurations in use by this environment.

environmentResourceDescription_triggers :: Lens' EnvironmentResourceDescription (Maybe [Trigger])

The AutoScaling triggers in use by this environment.

environmentResourceDescription_queues :: Lens' EnvironmentResourceDescription (Maybe [Queue])

The queues used by this environment.

environmentResourceDescription_instances :: Lens' EnvironmentResourceDescription (Maybe [Instance])

The Amazon EC2 instances used by this environment.

environmentResourceDescription_environmentName :: Lens' EnvironmentResourceDescription (Maybe Text)

The name of the environment.

environmentResourceDescription_loadBalancers :: Lens' EnvironmentResourceDescription (Maybe [LoadBalancer])

The LoadBalancers in use by this environment.

environmentResourceDescription_autoScalingGroups :: Lens' EnvironmentResourceDescription (Maybe [AutoScalingGroup])

The AutoScalingGroups used by this environment.