Network.AWS.ElasticBeanstalk.Types.EventDescription

Description

Documentation

data EventDescription

Describes an event.

See: newEventDescription smart constructor.

Constructors

EventDescription'

Fields

templateName :: Maybe Text

The name of the configuration associated with this event.

severity :: Maybe EventSeverity

The severity level of this event.

message :: Maybe Text

The event message.

eventDate :: Maybe ISO8601

The date when the event occurred.

platformArn :: Maybe Text

The ARN of the platform version.

environmentName :: Maybe Text

The name of the environment associated with this event.

versionLabel :: Maybe Text

The release label for the application version associated with this event.

requestId :: Maybe Text

The web service request ID for the activity of this event.

applicationName :: Maybe Text

The application associated with the event.

newEventDescription :: EventDescription

Create a value of EventDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:templateName:EventDescription', eventDescription_templateName - The name of the configuration associated with this event.

$sel:severity:EventDescription', eventDescription_severity - The severity level of this event.

$sel:message:EventDescription', eventDescription_message - The event message.

$sel:eventDate:EventDescription', eventDescription_eventDate - The date when the event occurred.

$sel:platformArn:EventDescription', eventDescription_platformArn - The ARN of the platform version.

$sel:environmentName:EventDescription', eventDescription_environmentName - The name of the environment associated with this event.

$sel:versionLabel:EventDescription', eventDescription_versionLabel - The release label for the application version associated with this event.

$sel:requestId:EventDescription', eventDescription_requestId - The web service request ID for the activity of this event.

$sel:applicationName:EventDescription', eventDescription_applicationName - The application associated with the event.

eventDescription_templateName :: Lens' EventDescription (Maybe Text)

The name of the configuration associated with this event.

eventDescription_severity :: Lens' EventDescription (Maybe EventSeverity)

The severity level of this event.

eventDescription_message :: Lens' EventDescription (Maybe Text)

The event message.

eventDescription_eventDate :: Lens' EventDescription (Maybe UTCTime)

The date when the event occurred.

eventDescription_platformArn :: Lens' EventDescription (Maybe Text)

The ARN of the platform version.

eventDescription_environmentName :: Lens' EventDescription (Maybe Text)

The name of the environment associated with this event.

eventDescription_versionLabel :: Lens' EventDescription (Maybe Text)

The release label for the application version associated with this event.

eventDescription_requestId :: Lens' EventDescription (Maybe Text)

The web service request ID for the activity of this event.

eventDescription_applicationName :: Lens' EventDescription (Maybe Text)

The application associated with the event.