Network.AWS.ElasticBeanstalk.Types.ManagedActionHistoryItem

Description

Documentation

data ManagedActionHistoryItem

The record of a completed or failed managed action.

See: newManagedActionHistoryItem smart constructor.

Constructors

ManagedActionHistoryItem'

Fields

status :: Maybe ActionHistoryStatus

The status of the action.

actionType :: Maybe ActionType

The type of the managed action.

executedTime :: Maybe ISO8601

The date and time that the action started executing.

actionId :: Maybe Text

A unique identifier for the managed action.

actionDescription :: Maybe Text

A description of the managed action.

finishedTime :: Maybe ISO8601

The date and time that the action finished executing.

failureDescription :: Maybe Text

If the action failed, a description of the failure.

failureType :: Maybe FailureType

If the action failed, the type of failure.

newManagedActionHistoryItem :: ManagedActionHistoryItem

Create a value of ManagedActionHistoryItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ManagedActionHistoryItem', managedActionHistoryItem_status - The status of the action.

$sel:actionType:ManagedActionHistoryItem', managedActionHistoryItem_actionType - The type of the managed action.

$sel:executedTime:ManagedActionHistoryItem', managedActionHistoryItem_executedTime - The date and time that the action started executing.

$sel:actionId:ManagedActionHistoryItem', managedActionHistoryItem_actionId - A unique identifier for the managed action.

$sel:actionDescription:ManagedActionHistoryItem', managedActionHistoryItem_actionDescription - A description of the managed action.

$sel:finishedTime:ManagedActionHistoryItem', managedActionHistoryItem_finishedTime - The date and time that the action finished executing.

$sel:failureDescription:ManagedActionHistoryItem', managedActionHistoryItem_failureDescription - If the action failed, a description of the failure.

$sel:failureType:ManagedActionHistoryItem', managedActionHistoryItem_failureType - If the action failed, the type of failure.

managedActionHistoryItem_status :: Lens' ManagedActionHistoryItem (Maybe ActionHistoryStatus)

The status of the action.

managedActionHistoryItem_actionType :: Lens' ManagedActionHistoryItem (Maybe ActionType)

The type of the managed action.

managedActionHistoryItem_executedTime :: Lens' ManagedActionHistoryItem (Maybe UTCTime)

The date and time that the action started executing.

managedActionHistoryItem_actionId :: Lens' ManagedActionHistoryItem (Maybe Text)

A unique identifier for the managed action.

managedActionHistoryItem_actionDescription :: Lens' ManagedActionHistoryItem (Maybe Text)

A description of the managed action.

managedActionHistoryItem_finishedTime :: Lens' ManagedActionHistoryItem (Maybe UTCTime)

The date and time that the action finished executing.

managedActionHistoryItem_failureDescription :: Lens' ManagedActionHistoryItem (Maybe Text)

If the action failed, a description of the failure.

managedActionHistoryItem_failureType :: Lens' ManagedActionHistoryItem (Maybe FailureType)

If the action failed, the type of failure.