Network.AWS.ElasticBeanstalk.Types.SingleInstanceHealth

Description

Documentation

data SingleInstanceHealth

Detailed health information about an Amazon EC2 instance in your Elastic Beanstalk environment.

See: newSingleInstanceHealth smart constructor.

Constructors

SingleInstanceHealth'

Fields

instanceId :: Maybe Text

The ID of the Amazon EC2 instance.

instanceType :: Maybe Text

The instance's type.

color :: Maybe Text

Represents the color indicator that gives you information about the health of the EC2 instance. For more information, see Health Colors and Statuses.

causes :: Maybe [Text]

Represents the causes, which provide more information about the current health status.

availabilityZone :: Maybe Text

The availability zone in which the instance runs.

launchedAt :: Maybe ISO8601

The time at which the EC2 instance was launched.

deployment :: Maybe Deployment

Information about the most recent deployment to an instance.

healthStatus :: Maybe Text

Returns the health status of the specified instance. For more information, see Health Colors and Statuses.

system :: Maybe SystemStatus

Operating system metrics from the instance.

applicationMetrics :: Maybe ApplicationMetrics

Request metrics from your application.

newSingleInstanceHealth :: SingleInstanceHealth

Create a value of SingleInstanceHealth with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceId:SingleInstanceHealth', singleInstanceHealth_instanceId - The ID of the Amazon EC2 instance.

$sel:instanceType:SingleInstanceHealth', singleInstanceHealth_instanceType - The instance's type.

$sel:color:SingleInstanceHealth', singleInstanceHealth_color - Represents the color indicator that gives you information about the health of the EC2 instance. For more information, see Health Colors and Statuses.

$sel:causes:SingleInstanceHealth', singleInstanceHealth_causes - Represents the causes, which provide more information about the current health status.

$sel:availabilityZone:SingleInstanceHealth', singleInstanceHealth_availabilityZone - The availability zone in which the instance runs.

$sel:launchedAt:SingleInstanceHealth', singleInstanceHealth_launchedAt - The time at which the EC2 instance was launched.

$sel:deployment:SingleInstanceHealth', singleInstanceHealth_deployment - Information about the most recent deployment to an instance.

$sel:healthStatus:SingleInstanceHealth', singleInstanceHealth_healthStatus - Returns the health status of the specified instance. For more information, see Health Colors and Statuses.

$sel:system:SingleInstanceHealth', singleInstanceHealth_system - Operating system metrics from the instance.

$sel:applicationMetrics:SingleInstanceHealth', singleInstanceHealth_applicationMetrics - Request metrics from your application.

singleInstanceHealth_instanceId :: Lens' SingleInstanceHealth (Maybe Text)

The ID of the Amazon EC2 instance.

singleInstanceHealth_instanceType :: Lens' SingleInstanceHealth (Maybe Text)

The instance's type.

singleInstanceHealth_color :: Lens' SingleInstanceHealth (Maybe Text)

Represents the color indicator that gives you information about the health of the EC2 instance. For more information, see Health Colors and Statuses.

singleInstanceHealth_causes :: Lens' SingleInstanceHealth (Maybe [Text])

Represents the causes, which provide more information about the current health status.

singleInstanceHealth_availabilityZone :: Lens' SingleInstanceHealth (Maybe Text)

The availability zone in which the instance runs.

singleInstanceHealth_launchedAt :: Lens' SingleInstanceHealth (Maybe UTCTime)

The time at which the EC2 instance was launched.

singleInstanceHealth_deployment :: Lens' SingleInstanceHealth (Maybe Deployment)

Information about the most recent deployment to an instance.

singleInstanceHealth_healthStatus :: Lens' SingleInstanceHealth (Maybe Text)

Returns the health status of the specified instance. For more information, see Health Colors and Statuses.

singleInstanceHealth_system :: Lens' SingleInstanceHealth (Maybe SystemStatus)

Operating system metrics from the instance.

singleInstanceHealth_applicationMetrics :: Lens' SingleInstanceHealth (Maybe ApplicationMetrics)

Request metrics from your application.