Network.AWS.ElasticBeanstalk.Types.SourceBuildInformation

Description

Documentation

data SourceBuildInformation

Location of the source code for an application version.

See: newSourceBuildInformation smart constructor.

Constructors

SourceBuildInformation'

Fields

sourceType :: SourceType

The type of repository.

Git

Zip

sourceRepository :: SourceRepository

Location where the repository is stored.

CodeCommit

S3

sourceLocation :: Text

The location of the source code, as a formatted string, depending on the value of SourceRepository

For CodeCommit, the format is the repository name and commit ID, separated by a forward slash. For example, my-git-repo/265cfa0cf6af46153527f55d6503ec030551f57a.

For S3, the format is the S3 bucket name and object key, separated by a forward slash. For example, my-s3-bucket/Folders/my-source-file.

newSourceBuildInformation

Arguments

:: SourceType
$sel:sourceType:SourceBuildInformation'

-> SourceRepository
$sel:sourceRepository:SourceBuildInformation'

-> Text
$sel:sourceLocation:SourceBuildInformation'

-> SourceBuildInformation

Create a value of SourceBuildInformation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceType:SourceBuildInformation', sourceBuildInformation_sourceType - The type of repository.

Git

Zip

$sel:sourceRepository:SourceBuildInformation', sourceBuildInformation_sourceRepository - Location where the repository is stored.

CodeCommit

S3

$sel:sourceLocation:SourceBuildInformation', sourceBuildInformation_sourceLocation - The location of the source code, as a formatted string, depending on the value of SourceRepository

For CodeCommit, the format is the repository name and commit ID, separated by a forward slash. For example, my-git-repo/265cfa0cf6af46153527f55d6503ec030551f57a.

For S3, the format is the S3 bucket name and object key, separated by a forward slash. For example, my-s3-bucket/Folders/my-source-file.

sourceBuildInformation_sourceType :: Lens' SourceBuildInformation SourceType

The type of repository.

Git

Zip

sourceBuildInformation_sourceRepository :: Lens' SourceBuildInformation SourceRepository

Location where the repository is stored.

CodeCommit

S3

sourceBuildInformation_sourceLocation :: Lens' SourceBuildInformation Text

The location of the source code, as a formatted string, depending on the value of SourceRepository

For CodeCommit, the format is the repository name and commit ID, separated by a forward slash. For example, my-git-repo/265cfa0cf6af46153527f55d6503ec030551f57a.

For S3, the format is the S3 bucket name and object key, separated by a forward slash. For example, my-s3-bucket/Folders/my-source-file.