Network.AWS.ElasticBeanstalk.UpdateApplication

Description

Updates the specified application to have the specified properties.

If a property (for example, description) is not provided, the value remains unchanged. To clear these properties, specify an empty string.

Creating a Request

data UpdateApplication

Request to update an application.

See: newUpdateApplication smart constructor.

Constructors

UpdateApplication'

Fields

description :: Maybe Text

A new description for the application.

Default: If not specified, AWS Elastic Beanstalk does not update the description.

applicationName :: Text

The name of the application to update. If no such application is found, UpdateApplication returns an InvalidParameterValue error.

newUpdateApplication

Arguments

:: Text
$sel:applicationName:UpdateApplication'

-> UpdateApplication

Create a value of UpdateApplication with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:UpdateApplication', updateApplication_description - A new description for the application.

Default: If not specified, AWS Elastic Beanstalk does not update the description.

$sel:applicationName:UpdateApplication', updateApplication_applicationName - The name of the application to update. If no such application is found, UpdateApplication returns an InvalidParameterValue error.

Request Lenses

updateApplication_description :: Lens' UpdateApplication (Maybe Text)

A new description for the application.

Default: If not specified, AWS Elastic Beanstalk does not update the description.

updateApplication_applicationName :: Lens' UpdateApplication Text

The name of the application to update. If no such application is found, UpdateApplication returns an InvalidParameterValue error.

Destructuring the Response

data ApplicationDescriptionMessage

Result message containing a single description of an application.

See: newApplicationDescriptionMessage smart constructor.

Constructors

ApplicationDescriptionMessage'

Fields

application :: Maybe ApplicationDescription

The ApplicationDescription of the application.

newApplicationDescriptionMessage :: ApplicationDescriptionMessage

Create a value of ApplicationDescriptionMessage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:application:ApplicationDescriptionMessage', applicationDescriptionMessage_application - The ApplicationDescription of the application.

Response Lenses

applicationDescriptionMessage_application :: Lens' ApplicationDescriptionMessage (Maybe ApplicationDescription)

The ApplicationDescription of the application.