Network.AWS.ElasticBeanstalk.UpdateEnvironment

Description

Updates the environment description, deploys a new application version, updates the configuration settings to an entirely new configuration template, or updates select configuration option values in the running environment.

Attempting to update both the release and configuration is not allowed and AWS Elastic Beanstalk returns an InvalidParameterCombination error.

When updating the configuration settings to a new template or individual settings, a draft configuration is created and DescribeConfigurationSettings for this environment returns two setting descriptions with different DeploymentStatus values.

Creating a Request

data UpdateEnvironment

Request to update an environment.

See: newUpdateEnvironment smart constructor.

Constructors

UpdateEnvironment'

Fields

templateName :: Maybe Text

If this parameter is specified, AWS Elastic Beanstalk deploys this configuration template to the environment. If no such configuration template is found, AWS Elastic Beanstalk returns an InvalidParameterValue error.

environmentId :: Maybe Text

The ID of the environment to update.

If no environment with this ID exists, AWS Elastic Beanstalk returns an InvalidParameterValue error.

Condition: You must specify either this or an EnvironmentName, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

solutionStackName :: Maybe Text

This specifies the platform version that the environment will run after the environment is updated.

groupName :: Maybe Text

The name of the group to which the target environment belongs. Specify a group name only if the environment's name is specified in an environment manifest and not with the environment name or environment ID parameters. See Environment Manifest (env.yaml) for details.

optionsToRemove :: Maybe [OptionSpecification]

A list of custom user-defined configuration options to remove from the configuration set for this environment.

platformArn :: Maybe Text

The ARN of the platform, if used.

environmentName :: Maybe Text

The name of the environment to update. If no environment with this name exists, AWS Elastic Beanstalk returns an InvalidParameterValue error.

Condition: You must specify either this or an EnvironmentId, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

versionLabel :: Maybe Text

If this parameter is specified, AWS Elastic Beanstalk deploys the named application version to the environment. If no such application version is found, returns an InvalidParameterValue error.

optionSettings :: Maybe [ConfigurationOptionSetting]

If specified, AWS Elastic Beanstalk updates the configuration set associated with the running environment and sets the specified configuration options to the requested value.

description :: Maybe Text

If this parameter is specified, AWS Elastic Beanstalk updates the description of this environment.

applicationName :: Maybe Text

The name of the application with which the environment is associated.

tier :: Maybe EnvironmentTier

This specifies the tier to use to update the environment.

Condition: At this time, if you change the tier version, name, or type, AWS Elastic Beanstalk returns InvalidParameterValue error.

newUpdateEnvironment :: UpdateEnvironment

Create a value of UpdateEnvironment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:templateName:UpdateEnvironment', updateEnvironment_templateName - If this parameter is specified, AWS Elastic Beanstalk deploys this configuration template to the environment. If no such configuration template is found, AWS Elastic Beanstalk returns an InvalidParameterValue error.

$sel:environmentId:UpdateEnvironment', updateEnvironment_environmentId - The ID of the environment to update.

If no environment with this ID exists, AWS Elastic Beanstalk returns an InvalidParameterValue error.

Condition: You must specify either this or an EnvironmentName, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

$sel:solutionStackName:UpdateEnvironment', updateEnvironment_solutionStackName - This specifies the platform version that the environment will run after the environment is updated.

$sel:groupName:UpdateEnvironment', updateEnvironment_groupName - The name of the group to which the target environment belongs. Specify a group name only if the environment's name is specified in an environment manifest and not with the environment name or environment ID parameters. See Environment Manifest (env.yaml) for details.

$sel:optionsToRemove:UpdateEnvironment', updateEnvironment_optionsToRemove - A list of custom user-defined configuration options to remove from the configuration set for this environment.

$sel:platformArn:UpdateEnvironment', updateEnvironment_platformArn - The ARN of the platform, if used.

$sel:environmentName:UpdateEnvironment', updateEnvironment_environmentName - The name of the environment to update. If no environment with this name exists, AWS Elastic Beanstalk returns an InvalidParameterValue error.

Condition: You must specify either this or an EnvironmentId, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

$sel:versionLabel:UpdateEnvironment', updateEnvironment_versionLabel - If this parameter is specified, AWS Elastic Beanstalk deploys the named application version to the environment. If no such application version is found, returns an InvalidParameterValue error.

$sel:optionSettings:UpdateEnvironment', updateEnvironment_optionSettings - If specified, AWS Elastic Beanstalk updates the configuration set associated with the running environment and sets the specified configuration options to the requested value.

$sel:description:UpdateEnvironment', updateEnvironment_description - If this parameter is specified, AWS Elastic Beanstalk updates the description of this environment.

$sel:applicationName:UpdateEnvironment', updateEnvironment_applicationName - The name of the application with which the environment is associated.

$sel:tier:UpdateEnvironment', updateEnvironment_tier - This specifies the tier to use to update the environment.

Condition: At this time, if you change the tier version, name, or type, AWS Elastic Beanstalk returns InvalidParameterValue error.

Request Lenses

updateEnvironment_templateName :: Lens' UpdateEnvironment (Maybe Text)

If this parameter is specified, AWS Elastic Beanstalk deploys this configuration template to the environment. If no such configuration template is found, AWS Elastic Beanstalk returns an InvalidParameterValue error.

updateEnvironment_environmentId :: Lens' UpdateEnvironment (Maybe Text)

The ID of the environment to update.

If no environment with this ID exists, AWS Elastic Beanstalk returns an InvalidParameterValue error.

Condition: You must specify either this or an EnvironmentName, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

updateEnvironment_solutionStackName :: Lens' UpdateEnvironment (Maybe Text)

This specifies the platform version that the environment will run after the environment is updated.

updateEnvironment_groupName :: Lens' UpdateEnvironment (Maybe Text)

The name of the group to which the target environment belongs. Specify a group name only if the environment's name is specified in an environment manifest and not with the environment name or environment ID parameters. See Environment Manifest (env.yaml) for details.

updateEnvironment_optionsToRemove :: Lens' UpdateEnvironment (Maybe [OptionSpecification])

A list of custom user-defined configuration options to remove from the configuration set for this environment.

updateEnvironment_platformArn :: Lens' UpdateEnvironment (Maybe Text)

The ARN of the platform, if used.

updateEnvironment_environmentName :: Lens' UpdateEnvironment (Maybe Text)

The name of the environment to update. If no environment with this name exists, AWS Elastic Beanstalk returns an InvalidParameterValue error.

Condition: You must specify either this or an EnvironmentId, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

updateEnvironment_versionLabel :: Lens' UpdateEnvironment (Maybe Text)

If this parameter is specified, AWS Elastic Beanstalk deploys the named application version to the environment. If no such application version is found, returns an InvalidParameterValue error.

updateEnvironment_optionSettings :: Lens' UpdateEnvironment (Maybe [ConfigurationOptionSetting])

If specified, AWS Elastic Beanstalk updates the configuration set associated with the running environment and sets the specified configuration options to the requested value.

updateEnvironment_description :: Lens' UpdateEnvironment (Maybe Text)

If this parameter is specified, AWS Elastic Beanstalk updates the description of this environment.

updateEnvironment_applicationName :: Lens' UpdateEnvironment (Maybe Text)

The name of the application with which the environment is associated.

updateEnvironment_tier :: Lens' UpdateEnvironment (Maybe EnvironmentTier)

This specifies the tier to use to update the environment.

Condition: At this time, if you change the tier version, name, or type, AWS Elastic Beanstalk returns InvalidParameterValue error.

Destructuring the Response

data EnvironmentDescription

Describes the properties of an environment.

See: newEnvironmentDescription smart constructor.

Constructors

EnvironmentDescription'

Fields

abortableOperationInProgress :: Maybe Bool

Indicates if there is an in-progress environment configuration update or application version deployment that you can cancel.

true: There is an update in progress.

false: There are no updates currently in progress.

templateName :: Maybe Text

The name of the configuration template used to originally launch this environment.

status :: Maybe EnvironmentStatus

The current operational status of the environment:

Launching: Environment is in the process of initial deployment.

Updating: Environment is in the process of updating its configuration settings or application version.

Ready: Environment is available to have an action performed on it, such as update or terminate.

Terminating: Environment is in the shut-down process.

Terminated: Environment is not running.

dateCreated :: Maybe ISO8601

The creation date for this environment.

environmentLinks :: Maybe [EnvironmentLink]

A list of links to other environments in the same group.

environmentId :: Maybe Text

The ID of this environment.

solutionStackName :: Maybe Text

The name of the SolutionStack deployed with this environment.

platformArn :: Maybe Text

The ARN of the platform version.

environmentName :: Maybe Text

The name of this environment.

health :: Maybe EnvironmentHealth

Describes the health status of the environment. AWS Elastic Beanstalk indicates the failure levels for a running environment:

Red: Indicates the environment is not responsive. Occurs when three or more consecutive failures occur for an environment.

Yellow: Indicates that something is wrong. Occurs when two consecutive failures occur for an environment.

Green: Indicates the environment is healthy and fully functional.

Grey: Default health for a new environment. The environment is not fully launched and health checks have not started or health checks are suspended during an UpdateEnvironment or RestartEnvironment request.

Default: Grey

versionLabel :: Maybe Text

The application version deployed in this environment.

cname :: Maybe Text

The URL to the CNAME for this environment.

dateUpdated :: Maybe ISO8601

The last modified date for this environment.

resources :: Maybe EnvironmentResourcesDescription

The description of the AWS resources used by this environment.

description :: Maybe Text

Describes this environment.

endpointURL :: Maybe Text

For load-balanced, autoscaling environments, the URL to the LoadBalancer. For single-instance environments, the IP address of the instance.

healthStatus :: Maybe EnvironmentHealthStatus

Returns the health status of the application running in your environment. For more information, see Health Colors and Statuses.

environmentArn :: Maybe Text

The environment's Amazon Resource Name (ARN), which can be used in other API requests that require an ARN.

applicationName :: Maybe Text

The name of the application associated with this environment.

tier :: Maybe EnvironmentTier

Describes the current tier of this environment.

operationsRole :: Maybe Text

The Amazon Resource Name (ARN) of the environment's operations role. For more information, see Operations roles in the AWS Elastic Beanstalk Developer Guide.

newEnvironmentDescription :: EnvironmentDescription

Create a value of EnvironmentDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:abortableOperationInProgress:EnvironmentDescription', environmentDescription_abortableOperationInProgress - Indicates if there is an in-progress environment configuration update or application version deployment that you can cancel.

true: There is an update in progress.

false: There are no updates currently in progress.

$sel:templateName:EnvironmentDescription', environmentDescription_templateName - The name of the configuration template used to originally launch this environment.

$sel:status:EnvironmentDescription', environmentDescription_status - The current operational status of the environment:

Launching: Environment is in the process of initial deployment.

Updating: Environment is in the process of updating its configuration settings or application version.

Ready: Environment is available to have an action performed on it, such as update or terminate.

Terminating: Environment is in the shut-down process.

Terminated: Environment is not running.

$sel:dateCreated:EnvironmentDescription', environmentDescription_dateCreated - The creation date for this environment.

$sel:environmentLinks:EnvironmentDescription', environmentDescription_environmentLinks - A list of links to other environments in the same group.

$sel:environmentId:EnvironmentDescription', environmentDescription_environmentId - The ID of this environment.

$sel:solutionStackName:EnvironmentDescription', environmentDescription_solutionStackName - The name of the SolutionStack deployed with this environment.

$sel:platformArn:EnvironmentDescription', environmentDescription_platformArn - The ARN of the platform version.

$sel:environmentName:EnvironmentDescription', environmentDescription_environmentName - The name of this environment.

$sel:health:EnvironmentDescription', environmentDescription_health - Describes the health status of the environment. AWS Elastic Beanstalk indicates the failure levels for a running environment:

Red: Indicates the environment is not responsive. Occurs when three or more consecutive failures occur for an environment.

Yellow: Indicates that something is wrong. Occurs when two consecutive failures occur for an environment.

Green: Indicates the environment is healthy and fully functional.

Grey: Default health for a new environment. The environment is not fully launched and health checks have not started or health checks are suspended during an UpdateEnvironment or RestartEnvironment request.

Default: Grey

$sel:versionLabel:EnvironmentDescription', environmentDescription_versionLabel - The application version deployed in this environment.

$sel:cname:EnvironmentDescription', environmentDescription_cname - The URL to the CNAME for this environment.

$sel:dateUpdated:EnvironmentDescription', environmentDescription_dateUpdated - The last modified date for this environment.

$sel:resources:EnvironmentDescription', environmentDescription_resources - The description of the AWS resources used by this environment.

$sel:description:EnvironmentDescription', environmentDescription_description - Describes this environment.

$sel:endpointURL:EnvironmentDescription', environmentDescription_endpointURL - For load-balanced, autoscaling environments, the URL to the LoadBalancer. For single-instance environments, the IP address of the instance.

$sel:healthStatus:EnvironmentDescription', environmentDescription_healthStatus - Returns the health status of the application running in your environment. For more information, see Health Colors and Statuses.

$sel:environmentArn:EnvironmentDescription', environmentDescription_environmentArn - The environment's Amazon Resource Name (ARN), which can be used in other API requests that require an ARN.

$sel:applicationName:EnvironmentDescription', environmentDescription_applicationName - The name of the application associated with this environment.

$sel:tier:EnvironmentDescription', environmentDescription_tier - Describes the current tier of this environment.

$sel:operationsRole:EnvironmentDescription', environmentDescription_operationsRole - The Amazon Resource Name (ARN) of the environment's operations role. For more information, see Operations roles in the AWS Elastic Beanstalk Developer Guide.

Response Lenses

environmentDescription_abortableOperationInProgress :: Lens' EnvironmentDescription (Maybe Bool)

Indicates if there is an in-progress environment configuration update or application version deployment that you can cancel.

true: There is an update in progress.

false: There are no updates currently in progress.

environmentDescription_templateName :: Lens' EnvironmentDescription (Maybe Text)

The name of the configuration template used to originally launch this environment.

environmentDescription_status :: Lens' EnvironmentDescription (Maybe EnvironmentStatus)

The current operational status of the environment:

Launching: Environment is in the process of initial deployment.

Updating: Environment is in the process of updating its configuration settings or application version.

Ready: Environment is available to have an action performed on it, such as update or terminate.

Terminating: Environment is in the shut-down process.

Terminated: Environment is not running.

environmentDescription_dateCreated :: Lens' EnvironmentDescription (Maybe UTCTime)

The creation date for this environment.

environmentDescription_environmentLinks :: Lens' EnvironmentDescription (Maybe [EnvironmentLink])

A list of links to other environments in the same group.

environmentDescription_environmentId :: Lens' EnvironmentDescription (Maybe Text)

The ID of this environment.

environmentDescription_solutionStackName :: Lens' EnvironmentDescription (Maybe Text)

The name of the SolutionStack deployed with this environment.

environmentDescription_platformArn :: Lens' EnvironmentDescription (Maybe Text)

The ARN of the platform version.

environmentDescription_environmentName :: Lens' EnvironmentDescription (Maybe Text)

The name of this environment.

environmentDescription_health :: Lens' EnvironmentDescription (Maybe EnvironmentHealth)

Describes the health status of the environment. AWS Elastic Beanstalk indicates the failure levels for a running environment:

Red: Indicates the environment is not responsive. Occurs when three or more consecutive failures occur for an environment.

Yellow: Indicates that something is wrong. Occurs when two consecutive failures occur for an environment.

Green: Indicates the environment is healthy and fully functional.

Grey: Default health for a new environment. The environment is not fully launched and health checks have not started or health checks are suspended during an UpdateEnvironment or RestartEnvironment request.

Default: Grey

environmentDescription_versionLabel :: Lens' EnvironmentDescription (Maybe Text)

The application version deployed in this environment.

environmentDescription_cname :: Lens' EnvironmentDescription (Maybe Text)

The URL to the CNAME for this environment.

environmentDescription_dateUpdated :: Lens' EnvironmentDescription (Maybe UTCTime)

The last modified date for this environment.

environmentDescription_resources :: Lens' EnvironmentDescription (Maybe EnvironmentResourcesDescription)

The description of the AWS resources used by this environment.

environmentDescription_description :: Lens' EnvironmentDescription (Maybe Text)

Describes this environment.

environmentDescription_endpointURL :: Lens' EnvironmentDescription (Maybe Text)

For load-balanced, autoscaling environments, the URL to the LoadBalancer. For single-instance environments, the IP address of the instance.

environmentDescription_healthStatus :: Lens' EnvironmentDescription (Maybe EnvironmentHealthStatus)

Returns the health status of the application running in your environment. For more information, see Health Colors and Statuses.

environmentDescription_environmentArn :: Lens' EnvironmentDescription (Maybe Text)

The environment's Amazon Resource Name (ARN), which can be used in other API requests that require an ARN.

environmentDescription_applicationName :: Lens' EnvironmentDescription (Maybe Text)

The name of the application associated with this environment.

environmentDescription_tier :: Lens' EnvironmentDescription (Maybe EnvironmentTier)

Describes the current tier of this environment.

environmentDescription_operationsRole :: Lens' EnvironmentDescription (Maybe Text)

The Amazon Resource Name (ARN) of the environment's operations role. For more information, see Operations roles in the AWS Elastic Beanstalk Developer Guide.