Network.AWS.ElasticBeanstalk.V2010_12_01.CreateApplication

Description

Creates an application that has one configuration template named default and no application versions. The &lt;code&gt;default&lt;/code&gt; configuration template is for a 32-bit version of the Amazon Linux operating system running the Tomcat 6 application container. &lt;/note&gt;"> https:/elasticbeanstalk.us-east-1.amazon.com?ApplicationName=SampleApp &Description=Sample%20Description &Operation=CreateApplication &AuthParams Sample Description SampleApp 2010-11-16T23:09:20.256Z 2010-11-16T23:09:20.256Z Default 8b00e053-f1d6-11df-8a78-9f77047e0d0c.

Documentation

createApplication

Arguments

:: Text
_camApplicationName

-> CreateApplication

Minimum specification for a CreateApplication request.

data CreateApplication

Constructors

CreateApplication

Fields

_camApplicationName :: Text

The name of the application. Constraint: This name must be unique within your account. If the specified name already exists, the action returns an InvalidParameterValue error.

_camDescription :: Maybe Text

Describes the application.

camDescription :: Lens' CreateApplication (Maybe Text)

camApplicationName :: Lens' CreateApplication Text

data CreateApplicationResponse

Constructors

CreateApplicationResponse

Fields

_adoApplication :: Maybe ApplicationDescription

The ApplicationDescription of the application.

adoApplication :: Iso' CreateApplicationResponse (Maybe ApplicationDescription)