Network.AWS.ElasticBeanstalk.V2010_12_01.DeleteApplicationVersion

Description

Deletes the specified version from the specified application. You cannot delete an application version that is associated with a running environment. https:/elasticbeanstalk.us-east-1.amazon.com?ApplicationName=SampleApp &VersionLabel=First%20Release &Operation=DeleteApplicationVersion &AuthParams 58dc7339-f272-11df-8a78-9f77047e0d0c.

Documentation

deleteApplicationVersion

Arguments

:: Text
_davmApplicationName

-> Text
_davmVersionLabel

-> DeleteApplicationVersion

Minimum specification for a DeleteApplicationVersion request.

data DeleteApplicationVersion

Constructors

DeleteApplicationVersion

Fields

_davmApplicationName :: Text

The name of the application to delete releases from.

_davmVersionLabel :: Text

The label of the version to delete.

_davmDeleteSourceBundle :: Maybe Bool

Indicates whether to delete the associated source bundle from Amazon S3: true: An attempt is made to delete the associated Amazon S3 source bundle specified at time of creation. false: No action is taken on the Amazon S3 source bundle specified at time of creation. Valid Values: true | false.

davmVersionLabel :: Lens' DeleteApplicationVersion Text

davmDeleteSourceBundle :: Lens' DeleteApplicationVersion (Maybe Bool)

davmApplicationName :: Lens' DeleteApplicationVersion Text

data DeleteApplicationVersionResponse

Constructors

DeleteApplicationVersionResponse