Network.AWS.ElasticBeanstalk.V2010_12_01.DeleteConfigurationTemplate

Description

Deletes the specified configuration template. When you launch an environment using a configuration template, the environment gets a copy of the template. You can delete or modify the environment's copy of the template without affecting the running environment. https:/elasticbeanstalk.us-east-1.amazon.com?ApplicationName=SampleApp &TemplateName=SampleAppTemplate &Operation=DeleteConfigurationTemplate &AuthParams af9cf1b6-f25e-11df-8a78-9f77047e0d0c.

Documentation

data DeleteConfigurationTemplate

Constructors

DeleteConfigurationTemplate

Fields

_dctmApplicationName :: Text

The name of the application to delete the configuration template from.

_dctmTemplateName :: Text

The name of the configuration template to delete.

dctmTemplateName :: Lens' DeleteConfigurationTemplate Text

dctmApplicationName :: Lens' DeleteConfigurationTemplate Text

data DeleteConfigurationTemplateResponse

Constructors

DeleteConfigurationTemplateResponse