Network.AWS.ElasticBeanstalk.V2010_12_01.DescribeEnvironments

Description

Returns descriptions for existing environments. https:/elasticbeanstalk.us-east-1.amazon.com?ApplicationName=SampleApp &IncludeDeleted=true &IncludedDeletedBackTo=2008-11-05T06%3A00%3A00Z &Operation=DescribeEnvironments &AuthParams Version1 Available SampleApp elasticbeanstalk-SampleApp-1394386994.us-east-1.elb.amazonaws.com SampleApp-jxb293wg7n.elasticbeanstalk.amazonaws.com Green e-icsgecu3wf 2010-11-17T04:01:40.668Z 32bit Amazon Linux running Tomcat 7 EnvDescrip SampleApp 2010-11-17T03:59:33.520Z 44790c68-f260-11df-8a78-9f77047e0d0c.

Documentation

describeEnvironments :: DescribeEnvironments

Minimum specification for a DescribeEnvironments request.

data DescribeEnvironments

Constructors

DescribeEnvironments

Fields

_denApplicationName :: Maybe Text

If specified, AWS Elastic Beanstalk restricts the returned descriptions to include only those that are associated with this application.

_denEnvironmentIds :: [Text]

If specified, AWS Elastic Beanstalk restricts the returned descriptions to include only those that have the specified IDs.

_denEnvironmentNames :: [Text]

If specified, AWS Elastic Beanstalk restricts the returned descriptions to include only those that have the specified names.

_denIncludeDeleted :: Maybe Bool

Indicates whether to include deleted environments: true: Environments that have been deleted after IncludedDeletedBackTo are displayed. false: Do not include deleted environments.

_denIncludedDeletedBackTo :: Maybe ISO8601

If specified when IncludeDeleted is set to true, then environments deleted after this date are displayed.

_denVersionLabel :: Maybe Text

If specified, AWS Elastic Beanstalk restricts the returned descriptions to include only those that are associated with this application version.

denVersionLabel :: Lens' DescribeEnvironments (Maybe Text)

denIncludedDeletedBackTo :: Lens' DescribeEnvironments (Maybe ISO8601)

denIncludeDeleted :: Lens' DescribeEnvironments (Maybe Bool)

denEnvironmentNames :: Lens' DescribeEnvironments [Text]

denEnvironmentIds :: Lens' DescribeEnvironments [Text]

denApplicationName :: Lens' DescribeEnvironments (Maybe Text)

data DescribeEnvironmentsResponse

Constructors

DescribeEnvironmentsResponse

Fields

_ednEnvironments :: [EnvironmentDescription]

Returns an EnvironmentDescription list.

ednEnvironments :: Iso' DescribeEnvironmentsResponse [EnvironmentDescription]