Network.AWS.ElasticBeanstalk.V2010_12_01.RequestEnvironmentInfo

Description

Initiates a request to compile the specified type of information of the deployed environment. Setting the InfoType to tail compiles the last lines from the application server log files of every Amazon EC2 instance in your environment. Use RetrieveEnvironmentInfo to access the compiled information. Related Topics RetrieveEnvironmentInfo https:/elasticbeanstalk.us-east-1.amazon.com?EnvironmentId=e-hc8mvnayrx &EnvironmentName=SampleAppVersion &InfoType=tail &Operation=RequestEnvironmentInfo &AuthParams 126a4ff3-f28a-11df-8a78-9f77047e0d0c.

Documentation

requestEnvironmentInfo

Arguments

:: EnvironmentInfoType
_reimInfoType

-> RequestEnvironmentInfo

Minimum specification for a RequestEnvironmentInfo request.

data RequestEnvironmentInfo

Constructors

RequestEnvironmentInfo

Fields

_reimInfoType :: EnvironmentInfoType

The type of information to request.

_reimEnvironmentId :: Maybe Text

The ID of the environment of the requested data. If no such environment is found, RequestEnvironmentInfo returns an InvalidParameterValue error. Condition: You must specify either this or an EnvironmentName, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

_reimEnvironmentName :: Maybe Text

The name of the environment of the requested data. If no such environment is found, RequestEnvironmentInfo returns an InvalidParameterValue error. Condition: You must specify either this or an EnvironmentId, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

reimInfoType :: Lens' RequestEnvironmentInfo EnvironmentInfoType

reimEnvironmentName :: Lens' RequestEnvironmentInfo (Maybe Text)

reimEnvironmentId :: Lens' RequestEnvironmentInfo (Maybe Text)

data RequestEnvironmentInfoResponse

Constructors

RequestEnvironmentInfoResponse