Network.AWS.ElasticBeanstalk.V2010_12_01.RestartAppServer

Description

Causes the environment to restart the application container server running on each Amazon EC2 instance. https:/elasticbeanstalk.us-east-1.amazon.com?EnvironmentId=e-hc8mvnayrx &EnvironmentName=SampleAppVersion &Operation=RestartAppServer &AuthParams 90e8d1d5-f28a-11df-8a78-9f77047e0d0c.

Documentation

restartAppServer :: RestartAppServer

Minimum specification for a RestartAppServer request.

data RestartAppServer

Constructors

RestartAppServer

Fields

_rasmEnvironmentId :: Maybe Text

The ID of the environment to restart the server for. Condition: You must specify either this or an EnvironmentName, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

_rasmEnvironmentName :: Maybe Text

The name of the environment to restart the server for. Condition: You must specify either this or an EnvironmentId, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

rasmEnvironmentName :: Lens' RestartAppServer (Maybe Text)

rasmEnvironmentId :: Lens' RestartAppServer (Maybe Text)

data RestartAppServerResponse

Constructors

RestartAppServerResponse