Network.AWS.ElasticBeanstalk.V2010_12_01.ValidateConfigurationSettings

Description

Takes a set of configuration settings and either a configuration template or environment, and determines whether those values are valid. This action returns a list of messages indicating any errors or warnings associated with the selection of option values. https:/elasticbeanstalk.us-east-1.amazon.com?ApplicationName=SampleApp &EnvironmentName=SampleAppVersion &OptionSettings.member.1.Namespace=aws%3Aautoscaling%3Atrigger &OptionSettings.member.1.OptionName=LowerThreshold &OptionSettings.member.1.Value=1000000 &Operation=ValidateConfigurationSettings &AuthParams 06f1cfff-f28f-11df-8a78-9f77047e0d0c.

Documentation

validateConfigurationSettings

Arguments

:: Text
_vcsmApplicationName

-> [ConfigurationOptionSetting]
_vcsmOptionSettings

-> ValidateConfigurationSettings

Minimum specification for a ValidateConfigurationSettings request.

data ValidateConfigurationSettings

Constructors

ValidateConfigurationSettings

Fields

_vcsmApplicationName :: Text

The name of the application that the configuration template or environment belongs to.

_vcsmOptionSettings :: [ConfigurationOptionSetting]

A list of the options and desired values to evaluate.

_vcsmTemplateName :: Maybe Text

The name of the configuration template to validate the settings against. Condition: You cannot specify both this and an environment name.

_vcsmEnvironmentName :: Maybe Text

The name of the environment to validate the settings against. Condition: You cannot specify both this and a configuration template name.

vcsmTemplateName :: Lens' ValidateConfigurationSettings (Maybe Text)

vcsmOptionSettings :: Lens' ValidateConfigurationSettings [ConfigurationOptionSetting]

vcsmEnvironmentName :: Lens' ValidateConfigurationSettings (Maybe Text)

vcsmApplicationName :: Lens' ValidateConfigurationSettings Text

data ValidateConfigurationSettingsResponse

Constructors

ValidateConfigurationSettingsResponse

Fields

_csvmMessages :: [ValidationMessage]

A list of ValidationMessage.

csvmMessages :: Iso' ValidateConfigurationSettingsResponse [ValidationMessage]