Network.AWS.ElasticInference.Lens

Description

Operations

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The ARN of the Elastic Inference Accelerator to list the tags for.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe (HashMap Text Text))

The tags of the Elastic Inference Accelerator.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

DescribeAcceleratorOfferings

describeAcceleratorOfferings_acceleratorTypes :: Lens' DescribeAcceleratorOfferings (Maybe [Text])

The list of accelerator types to describe.

describeAcceleratorOfferings_locationType :: Lens' DescribeAcceleratorOfferings LocationType

The location type that you want to describe accelerator type offerings for. It can assume the following values: region: will return the accelerator type offering at the regional level. availability-zone: will return the accelerator type offering at the availability zone level. availability-zone-id: will return the accelerator type offering at the availability zone level returning the availability zone id.

describeAcceleratorOfferingsResponse_acceleratorTypeOfferings :: Lens' DescribeAcceleratorOfferingsResponse (Maybe [AcceleratorTypeOffering])

The list of accelerator type offerings for a specific location.

describeAcceleratorOfferingsResponse_httpStatus :: Lens' DescribeAcceleratorOfferingsResponse Int

The response's http status code.

DescribeAccelerators

describeAccelerators_filters :: Lens' DescribeAccelerators (Maybe [Filter])

One or more filters. Filter names and values are case-sensitive. Valid filter names are: accelerator-types: can provide a list of accelerator type names to filter for. instance-id: can provide a list of EC2 instance ids to filter for.

describeAccelerators_nextToken :: Lens' DescribeAccelerators (Maybe Text)

A token to specify where to start paginating. This is the NextToken from a previously truncated response.

describeAccelerators_maxResults :: Lens' DescribeAccelerators (Maybe Natural)

The total number of items to return in the command's output. If the total number of items available is more than the value specified, a NextToken is provided in the command's output. To resume pagination, provide the NextToken value in the starting-token argument of a subsequent command. Do not use the NextToken response element directly outside of the AWS CLI.

describeAccelerators_acceleratorIds :: Lens' DescribeAccelerators (Maybe [Text])

The IDs of the accelerators to describe.

describeAcceleratorsResponse_acceleratorSet :: Lens' DescribeAcceleratorsResponse (Maybe [ElasticInferenceAccelerator])

The details of the Elastic Inference Accelerators.

describeAcceleratorsResponse_nextToken :: Lens' DescribeAcceleratorsResponse (Maybe Text)

A token to specify where to start paginating. This is the NextToken from a previously truncated response.

describeAcceleratorsResponse_httpStatus :: Lens' DescribeAcceleratorsResponse Int

The response's http status code.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The ARN of the Elastic Inference Accelerator to tag.

tagResource_tags :: Lens' TagResource (HashMap Text Text)

The tags to add to the Elastic Inference Accelerator.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

UntagResource

untagResource_resourceArn :: Lens' UntagResource Text

The ARN of the Elastic Inference Accelerator to untag.

untagResource_tagKeys :: Lens' UntagResource (NonEmpty Text)

The list of tags to remove from the Elastic Inference Accelerator.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

DescribeAcceleratorTypes

describeAcceleratorTypesResponse_acceleratorTypes :: Lens' DescribeAcceleratorTypesResponse (Maybe [AcceleratorType])

The available accelerator types.

describeAcceleratorTypesResponse_httpStatus :: Lens' DescribeAcceleratorTypesResponse Int

The response's http status code.

Types

AcceleratorType

acceleratorType_throughputInfo :: Lens' AcceleratorType (Maybe [KeyValuePair])

The throughput information of the Elastic Inference Accelerator type.

acceleratorType_memoryInfo :: Lens' AcceleratorType (Maybe MemoryInfo)

The memory information of the Elastic Inference Accelerator type.

acceleratorType_acceleratorTypeName :: Lens' AcceleratorType (Maybe Text)

The name of the Elastic Inference Accelerator type.

AcceleratorTypeOffering

acceleratorTypeOffering_acceleratorType :: Lens' AcceleratorTypeOffering (Maybe Text)

The name of the Elastic Inference Accelerator type.

acceleratorTypeOffering_location :: Lens' AcceleratorTypeOffering (Maybe Text)

The location for the offering. It will return either the region, availability zone or availability zone id for the offering depending on the locationType value.

acceleratorTypeOffering_locationType :: Lens' AcceleratorTypeOffering (Maybe LocationType)

The location type for the offering. It can assume the following values: region: defines that the offering is at the regional level. availability-zone: defines that the offering is at the availability zone level. availability-zone-id: defines that the offering is at the availability zone level, defined by the availability zone id.

ElasticInferenceAccelerator

elasticInferenceAccelerator_acceleratorType :: Lens' ElasticInferenceAccelerator (Maybe Text)

The type of the Elastic Inference Accelerator.

elasticInferenceAccelerator_acceleratorId :: Lens' ElasticInferenceAccelerator (Maybe Text)

The ID of the Elastic Inference Accelerator.

elasticInferenceAccelerator_attachedResource :: Lens' ElasticInferenceAccelerator (Maybe Text)

The ARN of the resource that the Elastic Inference Accelerator is attached to.

elasticInferenceAccelerator_acceleratorHealth :: Lens' ElasticInferenceAccelerator (Maybe ElasticInferenceAcceleratorHealth)

The health of the Elastic Inference Accelerator.

elasticInferenceAccelerator_availabilityZone :: Lens' ElasticInferenceAccelerator (Maybe Text)

The availability zone where the Elastic Inference Accelerator is present.

ElasticInferenceAcceleratorHealth

elasticInferenceAcceleratorHealth_status :: Lens' ElasticInferenceAcceleratorHealth (Maybe Text)

The health status of the Elastic Inference Accelerator.

Filter

filter_values :: Lens' Filter (Maybe [Text])

The values for the filter of the Elastic Inference Accelerator list.

filter_name :: Lens' Filter (Maybe Text)

The filter name for the Elastic Inference Accelerator list. It can assume the following values: accelerator-type: the type of Elastic Inference Accelerator to filter for. instance-id: an EC2 instance id to filter for.

KeyValuePair

keyValuePair_value :: Lens' KeyValuePair (Maybe Int)

The throughput value of the Elastic Inference Accelerator type.

keyValuePair_key :: Lens' KeyValuePair (Maybe Text)

The throughput value of the Elastic Inference Accelerator type. It can assume the following values: TFLOPS16bit: the throughput expressed in 16bit TeraFLOPS. TFLOPS32bit: the throughput expressed in 32bit TeraFLOPS.

MemoryInfo

memoryInfo_sizeInMiB :: Lens' MemoryInfo (Maybe Int)

The size in mebibytes of the Elastic Inference Accelerator type.