Network.AWS.ElasticInference.Types

Description

Service Configuration

defaultService :: Service

API version 2017-07-25 of the Amazon Elastic Inference SDK configuration.

Errors

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

Raised when an unexpected error occurred during request processing.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

Raised when the requested resource cannot be found.

_BadRequestException :: AsError a => Getting (First ServiceError) a ServiceError

Raised when a malformed input has been provided to the API.

LocationType

newtype LocationType

Constructors

LocationType'

Fields

fromLocationType :: Text

Bundled Patterns

pattern LocationType_Availability_zone :: LocationType

pattern LocationType_Availability_zone_id :: LocationType

pattern LocationType_Region :: LocationType

AcceleratorType

data AcceleratorType

The details of an Elastic Inference Accelerator type.

See: newAcceleratorType smart constructor.

Constructors

AcceleratorType'

Fields

throughputInfo :: Maybe [KeyValuePair]

The throughput information of the Elastic Inference Accelerator type.

memoryInfo :: Maybe MemoryInfo

The memory information of the Elastic Inference Accelerator type.

acceleratorTypeName :: Maybe Text

The name of the Elastic Inference Accelerator type.

newAcceleratorType :: AcceleratorType

Create a value of AcceleratorType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:throughputInfo:AcceleratorType', acceleratorType_throughputInfo - The throughput information of the Elastic Inference Accelerator type.

$sel:memoryInfo:AcceleratorType', acceleratorType_memoryInfo - The memory information of the Elastic Inference Accelerator type.

$sel:acceleratorTypeName:AcceleratorType', acceleratorType_acceleratorTypeName - The name of the Elastic Inference Accelerator type.

acceleratorType_throughputInfo :: Lens' AcceleratorType (Maybe [KeyValuePair])

The throughput information of the Elastic Inference Accelerator type.

acceleratorType_memoryInfo :: Lens' AcceleratorType (Maybe MemoryInfo)

The memory information of the Elastic Inference Accelerator type.

acceleratorType_acceleratorTypeName :: Lens' AcceleratorType (Maybe Text)

The name of the Elastic Inference Accelerator type.

AcceleratorTypeOffering

data AcceleratorTypeOffering

The offering for an Elastic Inference Accelerator type.

See: newAcceleratorTypeOffering smart constructor.

Constructors

AcceleratorTypeOffering'

Fields

acceleratorType :: Maybe Text

The name of the Elastic Inference Accelerator type.

location :: Maybe Text

The location for the offering. It will return either the region, availability zone or availability zone id for the offering depending on the locationType value.

locationType :: Maybe LocationType

The location type for the offering. It can assume the following values: region: defines that the offering is at the regional level. availability-zone: defines that the offering is at the availability zone level. availability-zone-id: defines that the offering is at the availability zone level, defined by the availability zone id.

newAcceleratorTypeOffering :: AcceleratorTypeOffering

Create a value of AcceleratorTypeOffering with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:acceleratorType:AcceleratorTypeOffering', acceleratorTypeOffering_acceleratorType - The name of the Elastic Inference Accelerator type.

$sel:location:AcceleratorTypeOffering', acceleratorTypeOffering_location - The location for the offering. It will return either the region, availability zone or availability zone id for the offering depending on the locationType value.

$sel:locationType:AcceleratorTypeOffering', acceleratorTypeOffering_locationType - The location type for the offering. It can assume the following values: region: defines that the offering is at the regional level. availability-zone: defines that the offering is at the availability zone level. availability-zone-id: defines that the offering is at the availability zone level, defined by the availability zone id.

acceleratorTypeOffering_acceleratorType :: Lens' AcceleratorTypeOffering (Maybe Text)

The name of the Elastic Inference Accelerator type.

acceleratorTypeOffering_location :: Lens' AcceleratorTypeOffering (Maybe Text)

The location for the offering. It will return either the region, availability zone or availability zone id for the offering depending on the locationType value.

acceleratorTypeOffering_locationType :: Lens' AcceleratorTypeOffering (Maybe LocationType)

The location type for the offering. It can assume the following values: region: defines that the offering is at the regional level. availability-zone: defines that the offering is at the availability zone level. availability-zone-id: defines that the offering is at the availability zone level, defined by the availability zone id.

ElasticInferenceAccelerator

data ElasticInferenceAccelerator

The details of an Elastic Inference Accelerator.

See: newElasticInferenceAccelerator smart constructor.

Constructors

ElasticInferenceAccelerator'

Fields

acceleratorType :: Maybe Text

The type of the Elastic Inference Accelerator.

acceleratorId :: Maybe Text

The ID of the Elastic Inference Accelerator.

attachedResource :: Maybe Text

The ARN of the resource that the Elastic Inference Accelerator is attached to.

acceleratorHealth :: Maybe ElasticInferenceAcceleratorHealth

The health of the Elastic Inference Accelerator.

availabilityZone :: Maybe Text

The availability zone where the Elastic Inference Accelerator is present.

newElasticInferenceAccelerator :: ElasticInferenceAccelerator

Create a value of ElasticInferenceAccelerator with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:acceleratorType:ElasticInferenceAccelerator', elasticInferenceAccelerator_acceleratorType - The type of the Elastic Inference Accelerator.

$sel:acceleratorId:ElasticInferenceAccelerator', elasticInferenceAccelerator_acceleratorId - The ID of the Elastic Inference Accelerator.

$sel:attachedResource:ElasticInferenceAccelerator', elasticInferenceAccelerator_attachedResource - The ARN of the resource that the Elastic Inference Accelerator is attached to.

$sel:acceleratorHealth:ElasticInferenceAccelerator', elasticInferenceAccelerator_acceleratorHealth - The health of the Elastic Inference Accelerator.

$sel:availabilityZone:ElasticInferenceAccelerator', elasticInferenceAccelerator_availabilityZone - The availability zone where the Elastic Inference Accelerator is present.

elasticInferenceAccelerator_acceleratorType :: Lens' ElasticInferenceAccelerator (Maybe Text)

The type of the Elastic Inference Accelerator.

elasticInferenceAccelerator_acceleratorId :: Lens' ElasticInferenceAccelerator (Maybe Text)

The ID of the Elastic Inference Accelerator.

elasticInferenceAccelerator_attachedResource :: Lens' ElasticInferenceAccelerator (Maybe Text)

The ARN of the resource that the Elastic Inference Accelerator is attached to.

elasticInferenceAccelerator_acceleratorHealth :: Lens' ElasticInferenceAccelerator (Maybe ElasticInferenceAcceleratorHealth)

The health of the Elastic Inference Accelerator.

elasticInferenceAccelerator_availabilityZone :: Lens' ElasticInferenceAccelerator (Maybe Text)

The availability zone where the Elastic Inference Accelerator is present.

ElasticInferenceAcceleratorHealth

data ElasticInferenceAcceleratorHealth

The health details of an Elastic Inference Accelerator.

See: newElasticInferenceAcceleratorHealth smart constructor.

Constructors

ElasticInferenceAcceleratorHealth'

Fields

status :: Maybe Text

The health status of the Elastic Inference Accelerator.

newElasticInferenceAcceleratorHealth :: ElasticInferenceAcceleratorHealth

Create a value of ElasticInferenceAcceleratorHealth with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ElasticInferenceAcceleratorHealth', elasticInferenceAcceleratorHealth_status - The health status of the Elastic Inference Accelerator.

elasticInferenceAcceleratorHealth_status :: Lens' ElasticInferenceAcceleratorHealth (Maybe Text)

The health status of the Elastic Inference Accelerator.

Filter

data Filter

A filter expression for the Elastic Inference Accelerator list.

See: newFilter smart constructor.

Constructors

Filter'

Fields

values :: Maybe [Text]

The values for the filter of the Elastic Inference Accelerator list.

name :: Maybe Text

The filter name for the Elastic Inference Accelerator list. It can assume the following values: accelerator-type: the type of Elastic Inference Accelerator to filter for. instance-id: an EC2 instance id to filter for.

newFilter :: Filter

Create a value of Filter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:values:Filter', filter_values - The values for the filter of the Elastic Inference Accelerator list.

$sel:name:Filter', filter_name - The filter name for the Elastic Inference Accelerator list. It can assume the following values: accelerator-type: the type of Elastic Inference Accelerator to filter for. instance-id: an EC2 instance id to filter for.

filter_values :: Lens' Filter (Maybe [Text])

The values for the filter of the Elastic Inference Accelerator list.

filter_name :: Lens' Filter (Maybe Text)

The filter name for the Elastic Inference Accelerator list. It can assume the following values: accelerator-type: the type of Elastic Inference Accelerator to filter for. instance-id: an EC2 instance id to filter for.

KeyValuePair

data KeyValuePair

A throughput entry for an Elastic Inference Accelerator type.

See: newKeyValuePair smart constructor.

Constructors

KeyValuePair'

Fields

value :: Maybe Int

The throughput value of the Elastic Inference Accelerator type.

key :: Maybe Text

The throughput value of the Elastic Inference Accelerator type. It can assume the following values: TFLOPS16bit: the throughput expressed in 16bit TeraFLOPS. TFLOPS32bit: the throughput expressed in 32bit TeraFLOPS.

newKeyValuePair :: KeyValuePair

Create a value of KeyValuePair with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:KeyValuePair', keyValuePair_value - The throughput value of the Elastic Inference Accelerator type.

$sel:key:KeyValuePair', keyValuePair_key - The throughput value of the Elastic Inference Accelerator type. It can assume the following values: TFLOPS16bit: the throughput expressed in 16bit TeraFLOPS. TFLOPS32bit: the throughput expressed in 32bit TeraFLOPS.

keyValuePair_value :: Lens' KeyValuePair (Maybe Int)

The throughput value of the Elastic Inference Accelerator type.

keyValuePair_key :: Lens' KeyValuePair (Maybe Text)

The throughput value of the Elastic Inference Accelerator type. It can assume the following values: TFLOPS16bit: the throughput expressed in 16bit TeraFLOPS. TFLOPS32bit: the throughput expressed in 32bit TeraFLOPS.

MemoryInfo

data MemoryInfo

The memory information of an Elastic Inference Accelerator type.

See: newMemoryInfo smart constructor.

Constructors

MemoryInfo'

Fields

sizeInMiB :: Maybe Int

The size in mebibytes of the Elastic Inference Accelerator type.

newMemoryInfo :: MemoryInfo

Create a value of MemoryInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sizeInMiB:MemoryInfo', memoryInfo_sizeInMiB - The size in mebibytes of the Elastic Inference Accelerator type.

memoryInfo_sizeInMiB :: Lens' MemoryInfo (Maybe Int)

The size in mebibytes of the Elastic Inference Accelerator type.