Network.AWS.ElasticInference.Types.AcceleratorTypeOffering

Description

Documentation

data AcceleratorTypeOffering

The offering for an Elastic Inference Accelerator type.

See: newAcceleratorTypeOffering smart constructor.

Constructors

AcceleratorTypeOffering'

Fields

acceleratorType :: Maybe Text

The name of the Elastic Inference Accelerator type.

location :: Maybe Text

The location for the offering. It will return either the region, availability zone or availability zone id for the offering depending on the locationType value.

locationType :: Maybe LocationType

The location type for the offering. It can assume the following values: region: defines that the offering is at the regional level. availability-zone: defines that the offering is at the availability zone level. availability-zone-id: defines that the offering is at the availability zone level, defined by the availability zone id.

newAcceleratorTypeOffering :: AcceleratorTypeOffering

Create a value of AcceleratorTypeOffering with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:acceleratorType:AcceleratorTypeOffering', acceleratorTypeOffering_acceleratorType - The name of the Elastic Inference Accelerator type.

$sel:location:AcceleratorTypeOffering', acceleratorTypeOffering_location - The location for the offering. It will return either the region, availability zone or availability zone id for the offering depending on the locationType value.

$sel:locationType:AcceleratorTypeOffering', acceleratorTypeOffering_locationType - The location type for the offering. It can assume the following values: region: defines that the offering is at the regional level. availability-zone: defines that the offering is at the availability zone level. availability-zone-id: defines that the offering is at the availability zone level, defined by the availability zone id.

acceleratorTypeOffering_acceleratorType :: Lens' AcceleratorTypeOffering (Maybe Text)

The name of the Elastic Inference Accelerator type.

acceleratorTypeOffering_location :: Lens' AcceleratorTypeOffering (Maybe Text)

The location for the offering. It will return either the region, availability zone or availability zone id for the offering depending on the locationType value.

acceleratorTypeOffering_locationType :: Lens' AcceleratorTypeOffering (Maybe LocationType)

The location type for the offering. It can assume the following values: region: defines that the offering is at the regional level. availability-zone: defines that the offering is at the availability zone level. availability-zone-id: defines that the offering is at the availability zone level, defined by the availability zone id.