Network.AWS.ElasticSearch.CancelElasticsearchServiceSoftwareUpdate

Description

Cancels a scheduled service software update for an Amazon ES domain. You can only perform this operation before the AutomatedUpdateDate and when the UpdateStatus is in the PENDING_UPDATE state.

Creating a Request

data CancelElasticsearchServiceSoftwareUpdate

Container for the parameters to the CancelElasticsearchServiceSoftwareUpdate operation. Specifies the name of the Elasticsearch domain that you wish to cancel a service software update on.

See: newCancelElasticsearchServiceSoftwareUpdate smart constructor.

Constructors

CancelElasticsearchServiceSoftwareUpdate'

Fields

domainName :: Text

The name of the domain that you want to stop the latest service software update on.

newCancelElasticsearchServiceSoftwareUpdate

Arguments

:: Text
$sel:domainName:CancelElasticsearchServiceSoftwareUpdate'

-> CancelElasticsearchServiceSoftwareUpdate

Create a value of CancelElasticsearchServiceSoftwareUpdate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainName:CancelElasticsearchServiceSoftwareUpdate', cancelElasticsearchServiceSoftwareUpdate_domainName - The name of the domain that you want to stop the latest service software update on.

Request Lenses

cancelElasticsearchServiceSoftwareUpdate_domainName :: Lens' CancelElasticsearchServiceSoftwareUpdate Text

The name of the domain that you want to stop the latest service software update on.

Destructuring the Response

data CancelElasticsearchServiceSoftwareUpdateResponse

The result of a CancelElasticsearchServiceSoftwareUpdate operation. Contains the status of the update.

See: newCancelElasticsearchServiceSoftwareUpdateResponse smart constructor.

Constructors

CancelElasticsearchServiceSoftwareUpdateResponse'

Fields

serviceSoftwareOptions :: Maybe ServiceSoftwareOptions

The current status of the Elasticsearch service software update.

httpStatus :: Int

The response's http status code.

newCancelElasticsearchServiceSoftwareUpdateResponse

Arguments

:: Int
$sel:httpStatus:CancelElasticsearchServiceSoftwareUpdateResponse'

-> CancelElasticsearchServiceSoftwareUpdateResponse

Create a value of CancelElasticsearchServiceSoftwareUpdateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceSoftwareOptions:CancelElasticsearchServiceSoftwareUpdateResponse', cancelElasticsearchServiceSoftwareUpdateResponse_serviceSoftwareOptions - The current status of the Elasticsearch service software update.

$sel:httpStatus:CancelElasticsearchServiceSoftwareUpdateResponse', cancelElasticsearchServiceSoftwareUpdateResponse_httpStatus - The response's http status code.

Response Lenses

cancelElasticsearchServiceSoftwareUpdateResponse_serviceSoftwareOptions :: Lens' CancelElasticsearchServiceSoftwareUpdateResponse (Maybe ServiceSoftwareOptions)

The current status of the Elasticsearch service software update.

cancelElasticsearchServiceSoftwareUpdateResponse_httpStatus :: Lens' CancelElasticsearchServiceSoftwareUpdateResponse Int

The response's http status code.