Network.AWS.ElasticSearch.CreateElasticsearchDomain

Description

Creates a new Elasticsearch domain. For more information, see Creating Elasticsearch Domains in the Amazon Elasticsearch Service Developer Guide.

Creating a Request

data CreateElasticsearchDomain

See: newCreateElasticsearchDomain smart constructor.

Constructors

CreateElasticsearchDomain'

Fields

eBSOptions :: Maybe EBSOptions

Options to enable, disable and specify the type and size of EBS storage volumes.

snapshotOptions :: Maybe SnapshotOptions

Option to set time, in UTC format, of the daily automated snapshot. Default value is 0 hours.

elasticsearchClusterConfig :: Maybe ElasticsearchClusterConfig

Configuration options for an Elasticsearch domain. Specifies the instance type and number of instances in the domain cluster.

domainEndpointOptions :: Maybe DomainEndpointOptions

Options to specify configuration that will be applied to the domain endpoint.

autoTuneOptions :: Maybe AutoTuneOptionsInput

Specifies Auto-Tune options.

vPCOptions :: Maybe VPCOptions

Options to specify the subnets and security groups for VPC endpoint. For more information, see Creating a VPC in VPC Endpoints for Amazon Elasticsearch Service Domains

accessPolicies :: Maybe Text

IAM access policy as a JSON-formatted string.

encryptionAtRestOptions :: Maybe EncryptionAtRestOptions

Specifies the Encryption At Rest Options.

cognitoOptions :: Maybe CognitoOptions

Options to specify the Cognito user and identity pools for Kibana authentication. For more information, see Amazon Cognito Authentication for Kibana.

nodeToNodeEncryptionOptions :: Maybe NodeToNodeEncryptionOptions

Specifies the NodeToNodeEncryptionOptions.

elasticsearchVersion :: Maybe Text

String of format X.Y to specify version for the Elasticsearch domain eg. "1.5" or "2.3". For more information, see Creating Elasticsearch Domains in the Amazon Elasticsearch Service Developer Guide.

advancedOptions :: Maybe (HashMap Text Text)

Option to allow references to indices in an HTTP request body. Must be false when configuring access to individual sub-resources. By default, the value is true. See Configuration Advanced Options for more information.

tagList :: Maybe [Tag]

A list of Tag added during domain creation.

advancedSecurityOptions :: Maybe AdvancedSecurityOptionsInput

Specifies advanced security options.

logPublishingOptions :: Maybe (HashMap LogType LogPublishingOption)

Map of LogType and LogPublishingOption, each containing options to publish a given type of Elasticsearch log.

domainName :: Text

The name of the Elasticsearch domain that you are creating. Domain names are unique across the domains owned by an account within an AWS region. Domain names must start with a lowercase letter and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

newCreateElasticsearchDomain

Arguments

:: Text
$sel:domainName:CreateElasticsearchDomain'

-> CreateElasticsearchDomain

Create a value of CreateElasticsearchDomain with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eBSOptions:CreateElasticsearchDomain', createElasticsearchDomain_eBSOptions - Options to enable, disable and specify the type and size of EBS storage volumes.

$sel:snapshotOptions:CreateElasticsearchDomain', createElasticsearchDomain_snapshotOptions - Option to set time, in UTC format, of the daily automated snapshot. Default value is 0 hours.

$sel:elasticsearchClusterConfig:CreateElasticsearchDomain', createElasticsearchDomain_elasticsearchClusterConfig - Configuration options for an Elasticsearch domain. Specifies the instance type and number of instances in the domain cluster.

$sel:domainEndpointOptions:CreateElasticsearchDomain', createElasticsearchDomain_domainEndpointOptions - Options to specify configuration that will be applied to the domain endpoint.

$sel:autoTuneOptions:CreateElasticsearchDomain', createElasticsearchDomain_autoTuneOptions - Specifies Auto-Tune options.

$sel:vPCOptions:CreateElasticsearchDomain', createElasticsearchDomain_vPCOptions - Options to specify the subnets and security groups for VPC endpoint. For more information, see Creating a VPC in VPC Endpoints for Amazon Elasticsearch Service Domains

$sel:accessPolicies:CreateElasticsearchDomain', createElasticsearchDomain_accessPolicies - IAM access policy as a JSON-formatted string.

$sel:encryptionAtRestOptions:CreateElasticsearchDomain', createElasticsearchDomain_encryptionAtRestOptions - Specifies the Encryption At Rest Options.

$sel:cognitoOptions:CreateElasticsearchDomain', createElasticsearchDomain_cognitoOptions - Options to specify the Cognito user and identity pools for Kibana authentication. For more information, see Amazon Cognito Authentication for Kibana.

$sel:nodeToNodeEncryptionOptions:CreateElasticsearchDomain', createElasticsearchDomain_nodeToNodeEncryptionOptions - Specifies the NodeToNodeEncryptionOptions.

$sel:elasticsearchVersion:CreateElasticsearchDomain', createElasticsearchDomain_elasticsearchVersion - String of format X.Y to specify version for the Elasticsearch domain eg. "1.5" or "2.3". For more information, see Creating Elasticsearch Domains in the Amazon Elasticsearch Service Developer Guide.

$sel:advancedOptions:CreateElasticsearchDomain', createElasticsearchDomain_advancedOptions - Option to allow references to indices in an HTTP request body. Must be false when configuring access to individual sub-resources. By default, the value is true. See Configuration Advanced Options for more information.

$sel:tagList:CreateElasticsearchDomain', createElasticsearchDomain_tagList - A list of Tag added during domain creation.

$sel:advancedSecurityOptions:CreateElasticsearchDomain', createElasticsearchDomain_advancedSecurityOptions - Specifies advanced security options.

$sel:logPublishingOptions:CreateElasticsearchDomain', createElasticsearchDomain_logPublishingOptions - Map of LogType and LogPublishingOption, each containing options to publish a given type of Elasticsearch log.

$sel:domainName:CreateElasticsearchDomain', createElasticsearchDomain_domainName - The name of the Elasticsearch domain that you are creating. Domain names are unique across the domains owned by an account within an AWS region. Domain names must start with a lowercase letter and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

Request Lenses

createElasticsearchDomain_eBSOptions :: Lens' CreateElasticsearchDomain (Maybe EBSOptions)

Options to enable, disable and specify the type and size of EBS storage volumes.

createElasticsearchDomain_snapshotOptions :: Lens' CreateElasticsearchDomain (Maybe SnapshotOptions)

Option to set time, in UTC format, of the daily automated snapshot. Default value is 0 hours.

createElasticsearchDomain_elasticsearchClusterConfig :: Lens' CreateElasticsearchDomain (Maybe ElasticsearchClusterConfig)

Configuration options for an Elasticsearch domain. Specifies the instance type and number of instances in the domain cluster.

createElasticsearchDomain_domainEndpointOptions :: Lens' CreateElasticsearchDomain (Maybe DomainEndpointOptions)

Options to specify configuration that will be applied to the domain endpoint.

createElasticsearchDomain_autoTuneOptions :: Lens' CreateElasticsearchDomain (Maybe AutoTuneOptionsInput)

Specifies Auto-Tune options.

createElasticsearchDomain_vPCOptions :: Lens' CreateElasticsearchDomain (Maybe VPCOptions)

Options to specify the subnets and security groups for VPC endpoint. For more information, see Creating a VPC in VPC Endpoints for Amazon Elasticsearch Service Domains

createElasticsearchDomain_accessPolicies :: Lens' CreateElasticsearchDomain (Maybe Text)

IAM access policy as a JSON-formatted string.

createElasticsearchDomain_encryptionAtRestOptions :: Lens' CreateElasticsearchDomain (Maybe EncryptionAtRestOptions)

Specifies the Encryption At Rest Options.

createElasticsearchDomain_cognitoOptions :: Lens' CreateElasticsearchDomain (Maybe CognitoOptions)

Options to specify the Cognito user and identity pools for Kibana authentication. For more information, see Amazon Cognito Authentication for Kibana.

createElasticsearchDomain_nodeToNodeEncryptionOptions :: Lens' CreateElasticsearchDomain (Maybe NodeToNodeEncryptionOptions)

Specifies the NodeToNodeEncryptionOptions.

createElasticsearchDomain_elasticsearchVersion :: Lens' CreateElasticsearchDomain (Maybe Text)

String of format X.Y to specify version for the Elasticsearch domain eg. "1.5" or "2.3". For more information, see Creating Elasticsearch Domains in the Amazon Elasticsearch Service Developer Guide.

createElasticsearchDomain_advancedOptions :: Lens' CreateElasticsearchDomain (Maybe (HashMap Text Text))

Option to allow references to indices in an HTTP request body. Must be false when configuring access to individual sub-resources. By default, the value is true. See Configuration Advanced Options for more information.

createElasticsearchDomain_tagList :: Lens' CreateElasticsearchDomain (Maybe [Tag])

A list of Tag added during domain creation.

createElasticsearchDomain_advancedSecurityOptions :: Lens' CreateElasticsearchDomain (Maybe AdvancedSecurityOptionsInput)

Specifies advanced security options.

createElasticsearchDomain_logPublishingOptions :: Lens' CreateElasticsearchDomain (Maybe (HashMap LogType LogPublishingOption))

Map of LogType and LogPublishingOption, each containing options to publish a given type of Elasticsearch log.

createElasticsearchDomain_domainName :: Lens' CreateElasticsearchDomain Text

The name of the Elasticsearch domain that you are creating. Domain names are unique across the domains owned by an account within an AWS region. Domain names must start with a lowercase letter and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

Destructuring the Response

data CreateElasticsearchDomainResponse

The result of a CreateElasticsearchDomain operation. Contains the status of the newly created Elasticsearch domain.

See: newCreateElasticsearchDomainResponse smart constructor.

Constructors

CreateElasticsearchDomainResponse'

Fields

domainStatus :: Maybe ElasticsearchDomainStatus

The status of the newly created Elasticsearch domain.

httpStatus :: Int

The response's http status code.

newCreateElasticsearchDomainResponse

Arguments

:: Int
$sel:httpStatus:CreateElasticsearchDomainResponse'

-> CreateElasticsearchDomainResponse

Create a value of CreateElasticsearchDomainResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainStatus:CreateElasticsearchDomainResponse', createElasticsearchDomainResponse_domainStatus - The status of the newly created Elasticsearch domain.

$sel:httpStatus:CreateElasticsearchDomainResponse', createElasticsearchDomainResponse_httpStatus - The response's http status code.

Response Lenses

createElasticsearchDomainResponse_domainStatus :: Lens' CreateElasticsearchDomainResponse (Maybe ElasticsearchDomainStatus)

The status of the newly created Elasticsearch domain.

createElasticsearchDomainResponse_httpStatus :: Lens' CreateElasticsearchDomainResponse Int

The response's http status code.