Network.AWS.ElasticSearch.DeleteInboundCrossClusterSearchConnection

Description

Allows the destination domain owner to delete an existing inbound cross-cluster search connection.

Creating a Request

data DeleteInboundCrossClusterSearchConnection

Container for the parameters to the DeleteInboundCrossClusterSearchConnection operation.

See: newDeleteInboundCrossClusterSearchConnection smart constructor.

Constructors

DeleteInboundCrossClusterSearchConnection'

Fields

crossClusterSearchConnectionId :: Text

The id of the inbound connection that you want to permanently delete.

newDeleteInboundCrossClusterSearchConnection

Arguments

:: Text
$sel:crossClusterSearchConnectionId:DeleteInboundCrossClusterSearchConnection'

-> DeleteInboundCrossClusterSearchConnection

Create a value of DeleteInboundCrossClusterSearchConnection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:crossClusterSearchConnectionId:DeleteInboundCrossClusterSearchConnection', deleteInboundCrossClusterSearchConnection_crossClusterSearchConnectionId - The id of the inbound connection that you want to permanently delete.

Request Lenses

deleteInboundCrossClusterSearchConnection_crossClusterSearchConnectionId :: Lens' DeleteInboundCrossClusterSearchConnection Text

The id of the inbound connection that you want to permanently delete.

Destructuring the Response

data DeleteInboundCrossClusterSearchConnectionResponse

The result of a DeleteInboundCrossClusterSearchConnection operation. Contains details of deleted inbound connection.

See: newDeleteInboundCrossClusterSearchConnectionResponse smart constructor.

Constructors

DeleteInboundCrossClusterSearchConnectionResponse'

Fields

crossClusterSearchConnection :: Maybe InboundCrossClusterSearchConnection

Specifies the InboundCrossClusterSearchConnection of deleted inbound connection.

httpStatus :: Int

The response's http status code.

newDeleteInboundCrossClusterSearchConnectionResponse

Arguments

:: Int
$sel:httpStatus:DeleteInboundCrossClusterSearchConnectionResponse'

-> DeleteInboundCrossClusterSearchConnectionResponse

Create a value of DeleteInboundCrossClusterSearchConnectionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:crossClusterSearchConnection:DeleteInboundCrossClusterSearchConnectionResponse', deleteInboundCrossClusterSearchConnectionResponse_crossClusterSearchConnection - Specifies the InboundCrossClusterSearchConnection of deleted inbound connection.

$sel:httpStatus:DeleteInboundCrossClusterSearchConnectionResponse', deleteInboundCrossClusterSearchConnectionResponse_httpStatus - The response's http status code.

Response Lenses

deleteInboundCrossClusterSearchConnectionResponse_crossClusterSearchConnection :: Lens' DeleteInboundCrossClusterSearchConnectionResponse (Maybe InboundCrossClusterSearchConnection)

Specifies the InboundCrossClusterSearchConnection of deleted inbound connection.

deleteInboundCrossClusterSearchConnectionResponse_httpStatus :: Lens' DeleteInboundCrossClusterSearchConnectionResponse Int

The response's http status code.