Network.AWS.ElasticSearch.DescribeElasticsearchDomains

Description

Returns domain configuration information about the specified Elasticsearch domains, including the domain ID, domain endpoint, and domain ARN.

Creating a Request

data DescribeElasticsearchDomains

Container for the parameters to the DescribeElasticsearchDomains operation. By default, the API returns the status of all Elasticsearch domains.

See: newDescribeElasticsearchDomains smart constructor.

Constructors

DescribeElasticsearchDomains'

Fields

domainNames :: [Text]

The Elasticsearch domains for which you want information.

newDescribeElasticsearchDomains :: DescribeElasticsearchDomains

Create a value of DescribeElasticsearchDomains with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainNames:DescribeElasticsearchDomains', describeElasticsearchDomains_domainNames - The Elasticsearch domains for which you want information.

Request Lenses

describeElasticsearchDomains_domainNames :: Lens' DescribeElasticsearchDomains [Text]

The Elasticsearch domains for which you want information.

Destructuring the Response

data DescribeElasticsearchDomainsResponse

The result of a DescribeElasticsearchDomains request. Contains the status of the specified domains or all domains owned by the account.

See: newDescribeElasticsearchDomainsResponse smart constructor.

Constructors

DescribeElasticsearchDomainsResponse'

Fields

httpStatus :: Int

The response's http status code.

domainStatusList :: [ElasticsearchDomainStatus]

The status of the domains requested in the DescribeElasticsearchDomains request.

newDescribeElasticsearchDomainsResponse

Arguments

:: Int
$sel:httpStatus:DescribeElasticsearchDomainsResponse'

-> DescribeElasticsearchDomainsResponse

Create a value of DescribeElasticsearchDomainsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DescribeElasticsearchDomainsResponse', describeElasticsearchDomainsResponse_httpStatus - The response's http status code.

$sel:domainStatusList:DescribeElasticsearchDomainsResponse', describeElasticsearchDomainsResponse_domainStatusList - The status of the domains requested in the DescribeElasticsearchDomains request.

Response Lenses

describeElasticsearchDomainsResponse_httpStatus :: Lens' DescribeElasticsearchDomainsResponse Int

The response's http status code.

describeElasticsearchDomainsResponse_domainStatusList :: Lens' DescribeElasticsearchDomainsResponse [ElasticsearchDomainStatus]

The status of the domains requested in the DescribeElasticsearchDomains request.