Network.AWS.ElasticSearch.DescribeReservedElasticsearchInstanceOfferings

Description

Lists available reserved Elasticsearch instance offerings.

This operation returns paginated results.

Creating a Request

data DescribeReservedElasticsearchInstanceOfferings

Container for parameters to DescribeReservedElasticsearchInstanceOfferings

See: newDescribeReservedElasticsearchInstanceOfferings smart constructor.

Constructors

DescribeReservedElasticsearchInstanceOfferings'

Fields

nextToken :: Maybe Text

NextToken should be sent in case if earlier API call produced result containing NextToken. It is used for pagination.

maxResults :: Maybe Int

Set this value to limit the number of results returned. If not specified, defaults to 100.

reservedElasticsearchInstanceOfferingId :: Maybe Text

The offering identifier filter value. Use this parameter to show only the available offering that matches the specified reservation identifier.

newDescribeReservedElasticsearchInstanceOfferings :: DescribeReservedElasticsearchInstanceOfferings

Create a value of DescribeReservedElasticsearchInstanceOfferings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeReservedElasticsearchInstanceOfferings', describeReservedElasticsearchInstanceOfferings_nextToken - NextToken should be sent in case if earlier API call produced result containing NextToken. It is used for pagination.

$sel:maxResults:DescribeReservedElasticsearchInstanceOfferings', describeReservedElasticsearchInstanceOfferings_maxResults - Set this value to limit the number of results returned. If not specified, defaults to 100.

$sel:reservedElasticsearchInstanceOfferingId:DescribeReservedElasticsearchInstanceOfferings', describeReservedElasticsearchInstanceOfferings_reservedElasticsearchInstanceOfferingId - The offering identifier filter value. Use this parameter to show only the available offering that matches the specified reservation identifier.

Request Lenses

describeReservedElasticsearchInstanceOfferings_nextToken :: Lens' DescribeReservedElasticsearchInstanceOfferings (Maybe Text)

NextToken should be sent in case if earlier API call produced result containing NextToken. It is used for pagination.

describeReservedElasticsearchInstanceOfferings_maxResults :: Lens' DescribeReservedElasticsearchInstanceOfferings (Maybe Int)

Set this value to limit the number of results returned. If not specified, defaults to 100.

describeReservedElasticsearchInstanceOfferings_reservedElasticsearchInstanceOfferingId :: Lens' DescribeReservedElasticsearchInstanceOfferings (Maybe Text)

The offering identifier filter value. Use this parameter to show only the available offering that matches the specified reservation identifier.

Destructuring the Response

data DescribeReservedElasticsearchInstanceOfferingsResponse

Container for results from DescribeReservedElasticsearchInstanceOfferings

See: newDescribeReservedElasticsearchInstanceOfferingsResponse smart constructor.

Constructors

DescribeReservedElasticsearchInstanceOfferingsResponse'

Fields

nextToken :: Maybe Text

Provides an identifier to allow retrieval of paginated results.

reservedElasticsearchInstanceOfferings :: Maybe [ReservedElasticsearchInstanceOffering]

List of reserved Elasticsearch instance offerings

httpStatus :: Int

The response's http status code.

newDescribeReservedElasticsearchInstanceOfferingsResponse

Arguments

:: Int
$sel:httpStatus:DescribeReservedElasticsearchInstanceOfferingsResponse'

-> DescribeReservedElasticsearchInstanceOfferingsResponse

Create a value of DescribeReservedElasticsearchInstanceOfferingsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeReservedElasticsearchInstanceOfferings', describeReservedElasticsearchInstanceOfferingsResponse_nextToken - Provides an identifier to allow retrieval of paginated results.

$sel:reservedElasticsearchInstanceOfferings:DescribeReservedElasticsearchInstanceOfferingsResponse', describeReservedElasticsearchInstanceOfferingsResponse_reservedElasticsearchInstanceOfferings - List of reserved Elasticsearch instance offerings

$sel:httpStatus:DescribeReservedElasticsearchInstanceOfferingsResponse', describeReservedElasticsearchInstanceOfferingsResponse_httpStatus - The response's http status code.

Response Lenses

describeReservedElasticsearchInstanceOfferingsResponse_nextToken :: Lens' DescribeReservedElasticsearchInstanceOfferingsResponse (Maybe Text)

Provides an identifier to allow retrieval of paginated results.

describeReservedElasticsearchInstanceOfferingsResponse_reservedElasticsearchInstanceOfferings :: Lens' DescribeReservedElasticsearchInstanceOfferingsResponse (Maybe [ReservedElasticsearchInstanceOffering])

List of reserved Elasticsearch instance offerings

describeReservedElasticsearchInstanceOfferingsResponse_httpStatus :: Lens' DescribeReservedElasticsearchInstanceOfferingsResponse Int

The response's http status code.