Network.AWS.ElasticSearch.DissociatePackage

Description

Dissociates a package from the Amazon ES domain.

Creating a Request

data DissociatePackage

Container for request parameters to DissociatePackage operation.

See: newDissociatePackage smart constructor.

Constructors

DissociatePackage'

Fields

packageID :: Text

Internal ID of the package that you want to associate with a domain. Use DescribePackages to find this value.

domainName :: Text

Name of the domain that you want to associate the package with.

newDissociatePackage

Arguments

:: Text
$sel:packageID:DissociatePackage'

-> Text
$sel:domainName:DissociatePackage'

-> DissociatePackage

Create a value of DissociatePackage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:packageID:DissociatePackage', dissociatePackage_packageID - Internal ID of the package that you want to associate with a domain. Use DescribePackages to find this value.

$sel:domainName:DissociatePackage', dissociatePackage_domainName - Name of the domain that you want to associate the package with.

Request Lenses

dissociatePackage_packageID :: Lens' DissociatePackage Text

Internal ID of the package that you want to associate with a domain. Use DescribePackages to find this value.

dissociatePackage_domainName :: Lens' DissociatePackage Text

Name of the domain that you want to associate the package with.

Destructuring the Response

data DissociatePackageResponse

Container for response returned by DissociatePackage operation.

See: newDissociatePackageResponse smart constructor.

Constructors

DissociatePackageResponse'

Fields

domainPackageDetails :: Maybe DomainPackageDetails

DomainPackageDetails

httpStatus :: Int

The response's http status code.

newDissociatePackageResponse

Arguments

:: Int
$sel:httpStatus:DissociatePackageResponse'

-> DissociatePackageResponse

Create a value of DissociatePackageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainPackageDetails:DissociatePackageResponse', dissociatePackageResponse_domainPackageDetails - DomainPackageDetails

$sel:httpStatus:DissociatePackageResponse', dissociatePackageResponse_httpStatus - The response's http status code.

Response Lenses

dissociatePackageResponse_domainPackageDetails :: Lens' DissociatePackageResponse (Maybe DomainPackageDetails)

DomainPackageDetails

dissociatePackageResponse_httpStatus :: Lens' DissociatePackageResponse Int

The response's http status code.