Network.AWS.ElasticSearch.GetPackageVersionHistory

Description

Returns a list of versions of the package, along with their creation time and commit message.

Creating a Request

data GetPackageVersionHistory

Container for request parameters to GetPackageVersionHistory operation.

See: newGetPackageVersionHistory smart constructor.

Constructors

GetPackageVersionHistory'

Fields

nextToken :: Maybe Text

Used for pagination. Only necessary if a previous API call includes a non-null NextToken value. If provided, returns results for the next page.

maxResults :: Maybe Int

Limits results to a maximum number of versions.

packageID :: Text

Returns an audit history of versions of the package.

newGetPackageVersionHistory

Arguments

:: Text
$sel:packageID:GetPackageVersionHistory'

-> GetPackageVersionHistory

Create a value of GetPackageVersionHistory with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetPackageVersionHistory', getPackageVersionHistory_nextToken - Used for pagination. Only necessary if a previous API call includes a non-null NextToken value. If provided, returns results for the next page.

$sel:maxResults:GetPackageVersionHistory', getPackageVersionHistory_maxResults - Limits results to a maximum number of versions.

$sel:packageID:GetPackageVersionHistory', getPackageVersionHistory_packageID - Returns an audit history of versions of the package.

Request Lenses

getPackageVersionHistory_nextToken :: Lens' GetPackageVersionHistory (Maybe Text)

Used for pagination. Only necessary if a previous API call includes a non-null NextToken value. If provided, returns results for the next page.

getPackageVersionHistory_maxResults :: Lens' GetPackageVersionHistory (Maybe Int)

Limits results to a maximum number of versions.

getPackageVersionHistory_packageID :: Lens' GetPackageVersionHistory Text

Returns an audit history of versions of the package.

Destructuring the Response

data GetPackageVersionHistoryResponse

Container for response returned by GetPackageVersionHistory operation.

See: newGetPackageVersionHistoryResponse smart constructor.

Constructors

GetPackageVersionHistoryResponse'

Fields

nextToken :: Maybe Text

packageID :: Maybe Text

packageVersionHistoryList :: Maybe [PackageVersionHistory]

List of PackageVersionHistory objects.

httpStatus :: Int

The response's http status code.

newGetPackageVersionHistoryResponse

Arguments

:: Int
$sel:httpStatus:GetPackageVersionHistoryResponse'

-> GetPackageVersionHistoryResponse

Create a value of GetPackageVersionHistoryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetPackageVersionHistory', getPackageVersionHistoryResponse_nextToken - Undocumented member.

$sel:packageID:GetPackageVersionHistory', getPackageVersionHistoryResponse_packageID - Undocumented member.

$sel:packageVersionHistoryList:GetPackageVersionHistoryResponse', getPackageVersionHistoryResponse_packageVersionHistoryList - List of PackageVersionHistory objects.

$sel:httpStatus:GetPackageVersionHistoryResponse', getPackageVersionHistoryResponse_httpStatus - The response's http status code.

Response Lenses

getPackageVersionHistoryResponse_nextToken :: Lens' GetPackageVersionHistoryResponse (Maybe Text)

Undocumented member.

getPackageVersionHistoryResponse_packageID :: Lens' GetPackageVersionHistoryResponse (Maybe Text)

Undocumented member.

getPackageVersionHistoryResponse_packageVersionHistoryList :: Lens' GetPackageVersionHistoryResponse (Maybe [PackageVersionHistory])

List of PackageVersionHistory objects.

getPackageVersionHistoryResponse_httpStatus :: Lens' GetPackageVersionHistoryResponse Int

The response's http status code.