Network.AWS.ElasticSearch.ListElasticsearchInstanceTypes

Description

List all Elasticsearch instance types that are supported for given ElasticsearchVersion

This operation returns paginated results.

Creating a Request

data ListElasticsearchInstanceTypes

Container for the parameters to the ListElasticsearchInstanceTypes operation.

See: newListElasticsearchInstanceTypes smart constructor.

Constructors

ListElasticsearchInstanceTypes'

Fields

nextToken :: Maybe Text

NextToken should be sent in case if earlier API call produced result containing NextToken. It is used for pagination.

domainName :: Maybe Text

DomainName represents the name of the Domain that we are trying to modify. This should be present only if we are querying for list of available Elasticsearch instance types when modifying existing domain.

maxResults :: Maybe Int

Set this value to limit the number of results returned. Value provided must be greater than 30 else it wont be honored.

elasticsearchVersion :: Text

Version of Elasticsearch for which list of supported elasticsearch instance types are needed.

newListElasticsearchInstanceTypes

Arguments

:: Text
$sel:elasticsearchVersion:ListElasticsearchInstanceTypes'

-> ListElasticsearchInstanceTypes

Create a value of ListElasticsearchInstanceTypes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListElasticsearchInstanceTypes', listElasticsearchInstanceTypes_nextToken - NextToken should be sent in case if earlier API call produced result containing NextToken. It is used for pagination.

$sel:domainName:ListElasticsearchInstanceTypes', listElasticsearchInstanceTypes_domainName - DomainName represents the name of the Domain that we are trying to modify. This should be present only if we are querying for list of available Elasticsearch instance types when modifying existing domain.

$sel:maxResults:ListElasticsearchInstanceTypes', listElasticsearchInstanceTypes_maxResults - Set this value to limit the number of results returned. Value provided must be greater than 30 else it wont be honored.

$sel:elasticsearchVersion:ListElasticsearchInstanceTypes', listElasticsearchInstanceTypes_elasticsearchVersion - Version of Elasticsearch for which list of supported elasticsearch instance types are needed.

Request Lenses

listElasticsearchInstanceTypes_nextToken :: Lens' ListElasticsearchInstanceTypes (Maybe Text)

NextToken should be sent in case if earlier API call produced result containing NextToken. It is used for pagination.

listElasticsearchInstanceTypes_domainName :: Lens' ListElasticsearchInstanceTypes (Maybe Text)

DomainName represents the name of the Domain that we are trying to modify. This should be present only if we are querying for list of available Elasticsearch instance types when modifying existing domain.

listElasticsearchInstanceTypes_maxResults :: Lens' ListElasticsearchInstanceTypes (Maybe Int)

Set this value to limit the number of results returned. Value provided must be greater than 30 else it wont be honored.

listElasticsearchInstanceTypes_elasticsearchVersion :: Lens' ListElasticsearchInstanceTypes Text

Version of Elasticsearch for which list of supported elasticsearch instance types are needed.

Destructuring the Response

data ListElasticsearchInstanceTypesResponse

Container for the parameters returned by ListElasticsearchInstanceTypes operation.

See: newListElasticsearchInstanceTypesResponse smart constructor.

Constructors

ListElasticsearchInstanceTypesResponse'

Fields

elasticsearchInstanceTypes :: Maybe [ESPartitionInstanceType]

List of instance types supported by Amazon Elasticsearch service for given ElasticsearchVersion

nextToken :: Maybe Text

In case if there are more results available NextToken would be present, make further request to the same API with received NextToken to paginate remaining results.

httpStatus :: Int

The response's http status code.

newListElasticsearchInstanceTypesResponse

Arguments

:: Int
$sel:httpStatus:ListElasticsearchInstanceTypesResponse'

-> ListElasticsearchInstanceTypesResponse

Create a value of ListElasticsearchInstanceTypesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:elasticsearchInstanceTypes:ListElasticsearchInstanceTypesResponse', listElasticsearchInstanceTypesResponse_elasticsearchInstanceTypes - List of instance types supported by Amazon Elasticsearch service for given ElasticsearchVersion

$sel:nextToken:ListElasticsearchInstanceTypes', listElasticsearchInstanceTypesResponse_nextToken - In case if there are more results available NextToken would be present, make further request to the same API with received NextToken to paginate remaining results.

$sel:httpStatus:ListElasticsearchInstanceTypesResponse', listElasticsearchInstanceTypesResponse_httpStatus - The response's http status code.

Response Lenses

listElasticsearchInstanceTypesResponse_elasticsearchInstanceTypes :: Lens' ListElasticsearchInstanceTypesResponse (Maybe [ESPartitionInstanceType])

List of instance types supported by Amazon Elasticsearch service for given ElasticsearchVersion

listElasticsearchInstanceTypesResponse_nextToken :: Lens' ListElasticsearchInstanceTypesResponse (Maybe Text)

In case if there are more results available NextToken would be present, make further request to the same API with received NextToken to paginate remaining results.

listElasticsearchInstanceTypesResponse_httpStatus :: Lens' ListElasticsearchInstanceTypesResponse Int

The response's http status code.