Network.AWS.ElasticSearch.Types.AdvancedOptionsStatus

Description

Documentation

data AdvancedOptionsStatus

Status of the advanced options for the specified Elasticsearch domain. Currently, the following advanced options are available:

Option to allow references to indices in an HTTP request body. Must be false when configuring access to individual sub-resources. By default, the value is true. See Configuration Advanced Options for more information.

Option to specify the percentage of heap space that is allocated to field data. By default, this setting is unbounded.

For more information, see Configuring Advanced Options.

See: newAdvancedOptionsStatus smart constructor.

Constructors

AdvancedOptionsStatus'

Fields

options :: HashMap Text Text

Specifies the status of advanced options for the specified Elasticsearch domain.

status :: OptionStatus

Specifies the status of OptionStatus for advanced options for the specified Elasticsearch domain.

newAdvancedOptionsStatus

Arguments

:: OptionStatus
$sel:status:AdvancedOptionsStatus'

-> AdvancedOptionsStatus

Create a value of AdvancedOptionsStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:options:AdvancedOptionsStatus', advancedOptionsStatus_options - Specifies the status of advanced options for the specified Elasticsearch domain.

$sel:status:AdvancedOptionsStatus', advancedOptionsStatus_status - Specifies the status of OptionStatus for advanced options for the specified Elasticsearch domain.

advancedOptionsStatus_options :: Lens' AdvancedOptionsStatus (HashMap Text Text)

Specifies the status of advanced options for the specified Elasticsearch domain.

advancedOptionsStatus_status :: Lens' AdvancedOptionsStatus OptionStatus

Specifies the status of OptionStatus for advanced options for the specified Elasticsearch domain.