Network.AWS.ElasticSearch.Types.EBSOptions

Description

Documentation

data EBSOptions

Options to enable, disable, and specify the properties of EBS storage volumes. For more information, see Configuring EBS-based Storage.

See: newEBSOptions smart constructor.

Constructors

EBSOptions'

Fields

eBSEnabled :: Maybe Bool

Specifies whether EBS-based storage is enabled.

volumeType :: Maybe VolumeType

Specifies the volume type for EBS-based storage.

iops :: Maybe Int

Specifies the IOPD for a Provisioned IOPS EBS volume (SSD).

volumeSize :: Maybe Int

Integer to specify the size of an EBS volume.

newEBSOptions :: EBSOptions

Create a value of EBSOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eBSEnabled:EBSOptions', eBSOptions_eBSEnabled - Specifies whether EBS-based storage is enabled.

$sel:volumeType:EBSOptions', eBSOptions_volumeType - Specifies the volume type for EBS-based storage.

$sel:iops:EBSOptions', eBSOptions_iops - Specifies the IOPD for a Provisioned IOPS EBS volume (SSD).

$sel:volumeSize:EBSOptions', eBSOptions_volumeSize - Integer to specify the size of an EBS volume.

eBSOptions_eBSEnabled :: Lens' EBSOptions (Maybe Bool)

Specifies whether EBS-based storage is enabled.

eBSOptions_volumeType :: Lens' EBSOptions (Maybe VolumeType)

Specifies the volume type for EBS-based storage.

eBSOptions_iops :: Lens' EBSOptions (Maybe Int)

Specifies the IOPD for a Provisioned IOPS EBS volume (SSD).

eBSOptions_volumeSize :: Lens' EBSOptions (Maybe Int)

Integer to specify the size of an EBS volume.