Network.AWS.ElasticSearch.Types.ElasticsearchClusterConfig

Description

Documentation

data ElasticsearchClusterConfig

Specifies the configuration for the domain cluster, such as the type and number of instances.

See: newElasticsearchClusterConfig smart constructor.

Constructors

ElasticsearchClusterConfig'

Fields

zoneAwarenessConfig :: Maybe ZoneAwarenessConfig

Specifies the zone awareness configuration for a domain when zone awareness is enabled.

dedicatedMasterCount :: Maybe Int

Total number of dedicated master nodes, active and on standby, for the cluster.

warmEnabled :: Maybe Bool

True to enable warm storage.

instanceType :: Maybe ESPartitionInstanceType

The instance type for an Elasticsearch cluster. UltraWarm instance types are not supported for data instances.

zoneAwarenessEnabled :: Maybe Bool

A boolean value to indicate whether zone awareness is enabled. See About Zone Awareness for more information.

dedicatedMasterEnabled :: Maybe Bool

A boolean value to indicate whether a dedicated master node is enabled. See About Dedicated Master Nodes for more information.

warmCount :: Maybe Int

The number of warm nodes in the cluster.

dedicatedMasterType :: Maybe ESPartitionInstanceType

The instance type for a dedicated master node.

warmType :: Maybe ESWarmPartitionInstanceType

The instance type for the Elasticsearch cluster's warm nodes.

coldStorageOptions :: Maybe ColdStorageOptions

Specifies the ColdStorageOptions config for Elasticsearch Domain

instanceCount :: Maybe Int

The number of instances in the specified domain cluster.

newElasticsearchClusterConfig :: ElasticsearchClusterConfig

Create a value of ElasticsearchClusterConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:zoneAwarenessConfig:ElasticsearchClusterConfig', elasticsearchClusterConfig_zoneAwarenessConfig - Specifies the zone awareness configuration for a domain when zone awareness is enabled.

$sel:dedicatedMasterCount:ElasticsearchClusterConfig', elasticsearchClusterConfig_dedicatedMasterCount - Total number of dedicated master nodes, active and on standby, for the cluster.

$sel:warmEnabled:ElasticsearchClusterConfig', elasticsearchClusterConfig_warmEnabled - True to enable warm storage.

$sel:instanceType:ElasticsearchClusterConfig', elasticsearchClusterConfig_instanceType - The instance type for an Elasticsearch cluster. UltraWarm instance types are not supported for data instances.

$sel:zoneAwarenessEnabled:ElasticsearchClusterConfig', elasticsearchClusterConfig_zoneAwarenessEnabled - A boolean value to indicate whether zone awareness is enabled. See About Zone Awareness for more information.

$sel:dedicatedMasterEnabled:ElasticsearchClusterConfig', elasticsearchClusterConfig_dedicatedMasterEnabled - A boolean value to indicate whether a dedicated master node is enabled. See About Dedicated Master Nodes for more information.

$sel:warmCount:ElasticsearchClusterConfig', elasticsearchClusterConfig_warmCount - The number of warm nodes in the cluster.

$sel:dedicatedMasterType:ElasticsearchClusterConfig', elasticsearchClusterConfig_dedicatedMasterType - The instance type for a dedicated master node.

$sel:warmType:ElasticsearchClusterConfig', elasticsearchClusterConfig_warmType - The instance type for the Elasticsearch cluster's warm nodes.

$sel:coldStorageOptions:ElasticsearchClusterConfig', elasticsearchClusterConfig_coldStorageOptions - Specifies the ColdStorageOptions config for Elasticsearch Domain

$sel:instanceCount:ElasticsearchClusterConfig', elasticsearchClusterConfig_instanceCount - The number of instances in the specified domain cluster.

elasticsearchClusterConfig_zoneAwarenessConfig :: Lens' ElasticsearchClusterConfig (Maybe ZoneAwarenessConfig)

Specifies the zone awareness configuration for a domain when zone awareness is enabled.

elasticsearchClusterConfig_dedicatedMasterCount :: Lens' ElasticsearchClusterConfig (Maybe Int)

Total number of dedicated master nodes, active and on standby, for the cluster.

elasticsearchClusterConfig_warmEnabled :: Lens' ElasticsearchClusterConfig (Maybe Bool)

True to enable warm storage.

elasticsearchClusterConfig_instanceType :: Lens' ElasticsearchClusterConfig (Maybe ESPartitionInstanceType)

The instance type for an Elasticsearch cluster. UltraWarm instance types are not supported for data instances.

elasticsearchClusterConfig_zoneAwarenessEnabled :: Lens' ElasticsearchClusterConfig (Maybe Bool)

A boolean value to indicate whether zone awareness is enabled. See About Zone Awareness for more information.

elasticsearchClusterConfig_dedicatedMasterEnabled :: Lens' ElasticsearchClusterConfig (Maybe Bool)

A boolean value to indicate whether a dedicated master node is enabled. See About Dedicated Master Nodes for more information.

elasticsearchClusterConfig_warmCount :: Lens' ElasticsearchClusterConfig (Maybe Int)

The number of warm nodes in the cluster.

elasticsearchClusterConfig_dedicatedMasterType :: Lens' ElasticsearchClusterConfig (Maybe ESPartitionInstanceType)

The instance type for a dedicated master node.

elasticsearchClusterConfig_warmType :: Lens' ElasticsearchClusterConfig (Maybe ESWarmPartitionInstanceType)

The instance type for the Elasticsearch cluster's warm nodes.

elasticsearchClusterConfig_coldStorageOptions :: Lens' ElasticsearchClusterConfig (Maybe ColdStorageOptions)

Specifies the ColdStorageOptions config for Elasticsearch Domain

elasticsearchClusterConfig_instanceCount :: Lens' ElasticsearchClusterConfig (Maybe Int)

The number of instances in the specified domain cluster.