Network.AWS.ElasticSearch.Types.ElasticsearchClusterConfigStatus

Description

Documentation

data ElasticsearchClusterConfigStatus

Specifies the configuration status for the specified Elasticsearch domain.

See: newElasticsearchClusterConfigStatus smart constructor.

Constructors

ElasticsearchClusterConfigStatus'

Fields

options :: ElasticsearchClusterConfig

Specifies the cluster configuration for the specified Elasticsearch domain.

status :: OptionStatus

Specifies the status of the configuration for the specified Elasticsearch domain.

newElasticsearchClusterConfigStatus

Arguments

:: ElasticsearchClusterConfig
$sel:options:ElasticsearchClusterConfigStatus'

-> OptionStatus
$sel:status:ElasticsearchClusterConfigStatus'

-> ElasticsearchClusterConfigStatus

Create a value of ElasticsearchClusterConfigStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:options:ElasticsearchClusterConfigStatus', elasticsearchClusterConfigStatus_options - Specifies the cluster configuration for the specified Elasticsearch domain.

$sel:status:ElasticsearchClusterConfigStatus', elasticsearchClusterConfigStatus_status - Specifies the status of the configuration for the specified Elasticsearch domain.

elasticsearchClusterConfigStatus_options :: Lens' ElasticsearchClusterConfigStatus ElasticsearchClusterConfig

Specifies the cluster configuration for the specified Elasticsearch domain.

elasticsearchClusterConfigStatus_status :: Lens' ElasticsearchClusterConfigStatus OptionStatus

Specifies the status of the configuration for the specified Elasticsearch domain.