Network.AWS.ElasticSearch.Types.ElasticsearchDomainConfig

Description

Documentation

data ElasticsearchDomainConfig

The configuration of an Elasticsearch domain.

See: newElasticsearchDomainConfig smart constructor.

Constructors

ElasticsearchDomainConfig'

Fields

eBSOptions :: Maybe EBSOptionsStatus

Specifies the EBSOptions for the Elasticsearch domain.

snapshotOptions :: Maybe SnapshotOptionsStatus

Specifies the SnapshotOptions for the Elasticsearch domain.

elasticsearchClusterConfig :: Maybe ElasticsearchClusterConfigStatus

Specifies the ElasticsearchClusterConfig for the Elasticsearch domain.

domainEndpointOptions :: Maybe DomainEndpointOptionsStatus

Specifies the DomainEndpointOptions for the Elasticsearch domain.

autoTuneOptions :: Maybe AutoTuneOptionsStatus

Specifies AutoTuneOptions for the domain.

vPCOptions :: Maybe VPCDerivedInfoStatus

The VPCOptions for the specified domain. For more information, see VPC Endpoints for Amazon Elasticsearch Service Domains.

accessPolicies :: Maybe AccessPoliciesStatus

IAM access policy as a JSON-formatted string.

encryptionAtRestOptions :: Maybe EncryptionAtRestOptionsStatus

Specifies the EncryptionAtRestOptions for the Elasticsearch domain.

cognitoOptions :: Maybe CognitoOptionsStatus

The CognitoOptions for the specified domain. For more information, see Amazon Cognito Authentication for Kibana.

nodeToNodeEncryptionOptions :: Maybe NodeToNodeEncryptionOptionsStatus

Specifies the NodeToNodeEncryptionOptions for the Elasticsearch domain.

elasticsearchVersion :: Maybe ElasticsearchVersionStatus

String of format X.Y to specify version for the Elasticsearch domain.

advancedOptions :: Maybe AdvancedOptionsStatus

Specifies the AdvancedOptions for the domain. See Configuring Advanced Options for more information.

advancedSecurityOptions :: Maybe AdvancedSecurityOptionsStatus

Specifies AdvancedSecurityOptions for the domain.

logPublishingOptions :: Maybe LogPublishingOptionsStatus

Log publishing options for the given domain.

newElasticsearchDomainConfig :: ElasticsearchDomainConfig

Create a value of ElasticsearchDomainConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eBSOptions:ElasticsearchDomainConfig', elasticsearchDomainConfig_eBSOptions - Specifies the EBSOptions for the Elasticsearch domain.

$sel:snapshotOptions:ElasticsearchDomainConfig', elasticsearchDomainConfig_snapshotOptions - Specifies the SnapshotOptions for the Elasticsearch domain.

$sel:elasticsearchClusterConfig:ElasticsearchDomainConfig', elasticsearchDomainConfig_elasticsearchClusterConfig - Specifies the ElasticsearchClusterConfig for the Elasticsearch domain.

$sel:domainEndpointOptions:ElasticsearchDomainConfig', elasticsearchDomainConfig_domainEndpointOptions - Specifies the DomainEndpointOptions for the Elasticsearch domain.

$sel:autoTuneOptions:ElasticsearchDomainConfig', elasticsearchDomainConfig_autoTuneOptions - Specifies AutoTuneOptions for the domain.

$sel:vPCOptions:ElasticsearchDomainConfig', elasticsearchDomainConfig_vPCOptions - The VPCOptions for the specified domain. For more information, see VPC Endpoints for Amazon Elasticsearch Service Domains.

$sel:accessPolicies:ElasticsearchDomainConfig', elasticsearchDomainConfig_accessPolicies - IAM access policy as a JSON-formatted string.

$sel:encryptionAtRestOptions:ElasticsearchDomainConfig', elasticsearchDomainConfig_encryptionAtRestOptions - Specifies the EncryptionAtRestOptions for the Elasticsearch domain.

$sel:cognitoOptions:ElasticsearchDomainConfig', elasticsearchDomainConfig_cognitoOptions - The CognitoOptions for the specified domain. For more information, see Amazon Cognito Authentication for Kibana.

$sel:nodeToNodeEncryptionOptions:ElasticsearchDomainConfig', elasticsearchDomainConfig_nodeToNodeEncryptionOptions - Specifies the NodeToNodeEncryptionOptions for the Elasticsearch domain.

$sel:elasticsearchVersion:ElasticsearchDomainConfig', elasticsearchDomainConfig_elasticsearchVersion - String of format X.Y to specify version for the Elasticsearch domain.

$sel:advancedOptions:ElasticsearchDomainConfig', elasticsearchDomainConfig_advancedOptions - Specifies the AdvancedOptions for the domain. See Configuring Advanced Options for more information.

$sel:advancedSecurityOptions:ElasticsearchDomainConfig', elasticsearchDomainConfig_advancedSecurityOptions - Specifies AdvancedSecurityOptions for the domain.

$sel:logPublishingOptions:ElasticsearchDomainConfig', elasticsearchDomainConfig_logPublishingOptions - Log publishing options for the given domain.

elasticsearchDomainConfig_eBSOptions :: Lens' ElasticsearchDomainConfig (Maybe EBSOptionsStatus)

Specifies the EBSOptions for the Elasticsearch domain.

elasticsearchDomainConfig_snapshotOptions :: Lens' ElasticsearchDomainConfig (Maybe SnapshotOptionsStatus)

Specifies the SnapshotOptions for the Elasticsearch domain.

elasticsearchDomainConfig_elasticsearchClusterConfig :: Lens' ElasticsearchDomainConfig (Maybe ElasticsearchClusterConfigStatus)

Specifies the ElasticsearchClusterConfig for the Elasticsearch domain.

elasticsearchDomainConfig_domainEndpointOptions :: Lens' ElasticsearchDomainConfig (Maybe DomainEndpointOptionsStatus)

Specifies the DomainEndpointOptions for the Elasticsearch domain.

elasticsearchDomainConfig_autoTuneOptions :: Lens' ElasticsearchDomainConfig (Maybe AutoTuneOptionsStatus)

Specifies AutoTuneOptions for the domain.

elasticsearchDomainConfig_vPCOptions :: Lens' ElasticsearchDomainConfig (Maybe VPCDerivedInfoStatus)

The VPCOptions for the specified domain. For more information, see VPC Endpoints for Amazon Elasticsearch Service Domains.

elasticsearchDomainConfig_accessPolicies :: Lens' ElasticsearchDomainConfig (Maybe AccessPoliciesStatus)

IAM access policy as a JSON-formatted string.

elasticsearchDomainConfig_encryptionAtRestOptions :: Lens' ElasticsearchDomainConfig (Maybe EncryptionAtRestOptionsStatus)

Specifies the EncryptionAtRestOptions for the Elasticsearch domain.

elasticsearchDomainConfig_cognitoOptions :: Lens' ElasticsearchDomainConfig (Maybe CognitoOptionsStatus)

The CognitoOptions for the specified domain. For more information, see Amazon Cognito Authentication for Kibana.

elasticsearchDomainConfig_nodeToNodeEncryptionOptions :: Lens' ElasticsearchDomainConfig (Maybe NodeToNodeEncryptionOptionsStatus)

Specifies the NodeToNodeEncryptionOptions for the Elasticsearch domain.

elasticsearchDomainConfig_elasticsearchVersion :: Lens' ElasticsearchDomainConfig (Maybe ElasticsearchVersionStatus)

String of format X.Y to specify version for the Elasticsearch domain.

elasticsearchDomainConfig_advancedOptions :: Lens' ElasticsearchDomainConfig (Maybe AdvancedOptionsStatus)

Specifies the AdvancedOptions for the domain. See Configuring Advanced Options for more information.

elasticsearchDomainConfig_advancedSecurityOptions :: Lens' ElasticsearchDomainConfig (Maybe AdvancedSecurityOptionsStatus)

Specifies AdvancedSecurityOptions for the domain.

elasticsearchDomainConfig_logPublishingOptions :: Lens' ElasticsearchDomainConfig (Maybe LogPublishingOptionsStatus)

Log publishing options for the given domain.