Network.AWS.ElasticSearch.Types.ElasticsearchDomainStatus

Description

Documentation

data ElasticsearchDomainStatus

The current status of an Elasticsearch domain.

See: newElasticsearchDomainStatus smart constructor.

Constructors

ElasticsearchDomainStatus'

Fields

eBSOptions :: Maybe EBSOptions

The EBSOptions for the specified domain. See Configuring EBS-based Storage for more information.

snapshotOptions :: Maybe SnapshotOptions

Specifies the status of the SnapshotOptions

domainEndpointOptions :: Maybe DomainEndpointOptions

The current status of the Elasticsearch domain's endpoint options.

upgradeProcessing :: Maybe Bool

The status of an Elasticsearch domain version upgrade. True if Amazon Elasticsearch Service is undergoing a version upgrade. False if the configuration is active.

endpoints :: Maybe (HashMap Text Text)

Map containing the Elasticsearch domain endpoints used to submit index and search requests. Example key, value: 'vpc','vpc-endpoint-h2dsd34efgyghrtguk5gt6j2foh4.us-east-1.es.amazonaws.com'.

autoTuneOptions :: Maybe AutoTuneOptionsOutput

The current status of the Elasticsearch domain's Auto-Tune options.

vPCOptions :: Maybe VPCDerivedInfo

The VPCOptions for the specified domain. For more information, see VPC Endpoints for Amazon Elasticsearch Service Domains.

accessPolicies :: Maybe Text

IAM access policy as a JSON-formatted string.

encryptionAtRestOptions :: Maybe EncryptionAtRestOptions

Specifies the status of the EncryptionAtRestOptions.

serviceSoftwareOptions :: Maybe ServiceSoftwareOptions

The current status of the Elasticsearch domain's service software.

cognitoOptions :: Maybe CognitoOptions

The CognitoOptions for the specified domain. For more information, see Amazon Cognito Authentication for Kibana.

nodeToNodeEncryptionOptions :: Maybe NodeToNodeEncryptionOptions

Specifies the status of the NodeToNodeEncryptionOptions.

elasticsearchVersion :: Maybe Text

advancedOptions :: Maybe (HashMap Text Text)

Specifies the status of the AdvancedOptions

endpoint :: Maybe Text

The Elasticsearch domain endpoint that you use to submit index and search requests.

processing :: Maybe Bool

The status of the Elasticsearch domain configuration. True if Amazon Elasticsearch Service is processing configuration changes. False if the configuration is active.

created :: Maybe Bool

The domain creation status. True if the creation of an Elasticsearch domain is complete. False if domain creation is still in progress.

advancedSecurityOptions :: Maybe AdvancedSecurityOptions

The current status of the Elasticsearch domain's advanced security options.

logPublishingOptions :: Maybe (HashMap LogType LogPublishingOption)

Log publishing options for the given domain.

deleted :: Maybe Bool

The domain deletion status. True if a delete request has been received for the domain but resource cleanup is still in progress. False if the domain has not been deleted. Once domain deletion is complete, the status of the domain is no longer returned.

domainId :: Text

The unique identifier for the specified Elasticsearch domain.

domainName :: Text

The name of an Elasticsearch domain. Domain names are unique across the domains owned by an account within an AWS region. Domain names start with a letter or number and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

arn :: Text

The Amazon resource name (ARN) of an Elasticsearch domain. See Identifiers for IAM Entities in Using AWS Identity and Access Management for more information.

elasticsearchClusterConfig :: ElasticsearchClusterConfig

The type and number of instances in the domain cluster.

newElasticsearchDomainStatus

Arguments

:: Text
$sel:domainId:ElasticsearchDomainStatus'

-> Text
$sel:domainName:ElasticsearchDomainStatus'

-> Text
$sel:arn:ElasticsearchDomainStatus'

-> ElasticsearchClusterConfig
$sel:elasticsearchClusterConfig:ElasticsearchDomainStatus'

-> ElasticsearchDomainStatus

Create a value of ElasticsearchDomainStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eBSOptions:ElasticsearchDomainStatus', elasticsearchDomainStatus_eBSOptions - The EBSOptions for the specified domain. See Configuring EBS-based Storage for more information.

$sel:snapshotOptions:ElasticsearchDomainStatus', elasticsearchDomainStatus_snapshotOptions - Specifies the status of the SnapshotOptions

$sel:domainEndpointOptions:ElasticsearchDomainStatus', elasticsearchDomainStatus_domainEndpointOptions - The current status of the Elasticsearch domain's endpoint options.

$sel:upgradeProcessing:ElasticsearchDomainStatus', elasticsearchDomainStatus_upgradeProcessing - The status of an Elasticsearch domain version upgrade. True if Amazon Elasticsearch Service is undergoing a version upgrade. False if the configuration is active.

$sel:endpoints:ElasticsearchDomainStatus', elasticsearchDomainStatus_endpoints - Map containing the Elasticsearch domain endpoints used to submit index and search requests. Example key, value: 'vpc','vpc-endpoint-h2dsd34efgyghrtguk5gt6j2foh4.us-east-1.es.amazonaws.com'.

$sel:autoTuneOptions:ElasticsearchDomainStatus', elasticsearchDomainStatus_autoTuneOptions - The current status of the Elasticsearch domain's Auto-Tune options.

$sel:vPCOptions:ElasticsearchDomainStatus', elasticsearchDomainStatus_vPCOptions - The VPCOptions for the specified domain. For more information, see VPC Endpoints for Amazon Elasticsearch Service Domains.

$sel:accessPolicies:ElasticsearchDomainStatus', elasticsearchDomainStatus_accessPolicies - IAM access policy as a JSON-formatted string.

$sel:encryptionAtRestOptions:ElasticsearchDomainStatus', elasticsearchDomainStatus_encryptionAtRestOptions - Specifies the status of the EncryptionAtRestOptions.

$sel:serviceSoftwareOptions:ElasticsearchDomainStatus', elasticsearchDomainStatus_serviceSoftwareOptions - The current status of the Elasticsearch domain's service software.

$sel:cognitoOptions:ElasticsearchDomainStatus', elasticsearchDomainStatus_cognitoOptions - The CognitoOptions for the specified domain. For more information, see Amazon Cognito Authentication for Kibana.

$sel:nodeToNodeEncryptionOptions:ElasticsearchDomainStatus', elasticsearchDomainStatus_nodeToNodeEncryptionOptions - Specifies the status of the NodeToNodeEncryptionOptions.

$sel:elasticsearchVersion:ElasticsearchDomainStatus', elasticsearchDomainStatus_elasticsearchVersion - Undocumented member.

$sel:advancedOptions:ElasticsearchDomainStatus', elasticsearchDomainStatus_advancedOptions - Specifies the status of the AdvancedOptions

$sel:endpoint:ElasticsearchDomainStatus', elasticsearchDomainStatus_endpoint - The Elasticsearch domain endpoint that you use to submit index and search requests.

$sel:processing:ElasticsearchDomainStatus', elasticsearchDomainStatus_processing - The status of the Elasticsearch domain configuration. True if Amazon Elasticsearch Service is processing configuration changes. False if the configuration is active.

$sel:created:ElasticsearchDomainStatus', elasticsearchDomainStatus_created - The domain creation status. True if the creation of an Elasticsearch domain is complete. False if domain creation is still in progress.

$sel:advancedSecurityOptions:ElasticsearchDomainStatus', elasticsearchDomainStatus_advancedSecurityOptions - The current status of the Elasticsearch domain's advanced security options.

$sel:logPublishingOptions:ElasticsearchDomainStatus', elasticsearchDomainStatus_logPublishingOptions - Log publishing options for the given domain.

$sel:deleted:ElasticsearchDomainStatus', elasticsearchDomainStatus_deleted - The domain deletion status. True if a delete request has been received for the domain but resource cleanup is still in progress. False if the domain has not been deleted. Once domain deletion is complete, the status of the domain is no longer returned.

$sel:domainId:ElasticsearchDomainStatus', elasticsearchDomainStatus_domainId - The unique identifier for the specified Elasticsearch domain.

$sel:domainName:ElasticsearchDomainStatus', elasticsearchDomainStatus_domainName - The name of an Elasticsearch domain. Domain names are unique across the domains owned by an account within an AWS region. Domain names start with a letter or number and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

$sel:arn:ElasticsearchDomainStatus', elasticsearchDomainStatus_arn - The Amazon resource name (ARN) of an Elasticsearch domain. See Identifiers for IAM Entities in Using AWS Identity and Access Management for more information.

$sel:elasticsearchClusterConfig:ElasticsearchDomainStatus', elasticsearchDomainStatus_elasticsearchClusterConfig - The type and number of instances in the domain cluster.

elasticsearchDomainStatus_eBSOptions :: Lens' ElasticsearchDomainStatus (Maybe EBSOptions)

The EBSOptions for the specified domain. See Configuring EBS-based Storage for more information.

elasticsearchDomainStatus_snapshotOptions :: Lens' ElasticsearchDomainStatus (Maybe SnapshotOptions)

Specifies the status of the SnapshotOptions

elasticsearchDomainStatus_domainEndpointOptions :: Lens' ElasticsearchDomainStatus (Maybe DomainEndpointOptions)

The current status of the Elasticsearch domain's endpoint options.

elasticsearchDomainStatus_upgradeProcessing :: Lens' ElasticsearchDomainStatus (Maybe Bool)

The status of an Elasticsearch domain version upgrade. True if Amazon Elasticsearch Service is undergoing a version upgrade. False if the configuration is active.

elasticsearchDomainStatus_endpoints :: Lens' ElasticsearchDomainStatus (Maybe (HashMap Text Text))

Map containing the Elasticsearch domain endpoints used to submit index and search requests. Example key, value: 'vpc','vpc-endpoint-h2dsd34efgyghrtguk5gt6j2foh4.us-east-1.es.amazonaws.com'.

elasticsearchDomainStatus_autoTuneOptions :: Lens' ElasticsearchDomainStatus (Maybe AutoTuneOptionsOutput)

The current status of the Elasticsearch domain's Auto-Tune options.

elasticsearchDomainStatus_vPCOptions :: Lens' ElasticsearchDomainStatus (Maybe VPCDerivedInfo)

The VPCOptions for the specified domain. For more information, see VPC Endpoints for Amazon Elasticsearch Service Domains.

elasticsearchDomainStatus_accessPolicies :: Lens' ElasticsearchDomainStatus (Maybe Text)

IAM access policy as a JSON-formatted string.

elasticsearchDomainStatus_encryptionAtRestOptions :: Lens' ElasticsearchDomainStatus (Maybe EncryptionAtRestOptions)

Specifies the status of the EncryptionAtRestOptions.

elasticsearchDomainStatus_serviceSoftwareOptions :: Lens' ElasticsearchDomainStatus (Maybe ServiceSoftwareOptions)

The current status of the Elasticsearch domain's service software.

elasticsearchDomainStatus_cognitoOptions :: Lens' ElasticsearchDomainStatus (Maybe CognitoOptions)

The CognitoOptions for the specified domain. For more information, see Amazon Cognito Authentication for Kibana.

elasticsearchDomainStatus_nodeToNodeEncryptionOptions :: Lens' ElasticsearchDomainStatus (Maybe NodeToNodeEncryptionOptions)

Specifies the status of the NodeToNodeEncryptionOptions.

elasticsearchDomainStatus_elasticsearchVersion :: Lens' ElasticsearchDomainStatus (Maybe Text)

Undocumented member.

elasticsearchDomainStatus_advancedOptions :: Lens' ElasticsearchDomainStatus (Maybe (HashMap Text Text))

Specifies the status of the AdvancedOptions

elasticsearchDomainStatus_endpoint :: Lens' ElasticsearchDomainStatus (Maybe Text)

The Elasticsearch domain endpoint that you use to submit index and search requests.

elasticsearchDomainStatus_processing :: Lens' ElasticsearchDomainStatus (Maybe Bool)

The status of the Elasticsearch domain configuration. True if Amazon Elasticsearch Service is processing configuration changes. False if the configuration is active.

elasticsearchDomainStatus_created :: Lens' ElasticsearchDomainStatus (Maybe Bool)

The domain creation status. True if the creation of an Elasticsearch domain is complete. False if domain creation is still in progress.

elasticsearchDomainStatus_advancedSecurityOptions :: Lens' ElasticsearchDomainStatus (Maybe AdvancedSecurityOptions)

The current status of the Elasticsearch domain's advanced security options.

elasticsearchDomainStatus_logPublishingOptions :: Lens' ElasticsearchDomainStatus (Maybe (HashMap LogType LogPublishingOption))

Log publishing options for the given domain.

elasticsearchDomainStatus_deleted :: Lens' ElasticsearchDomainStatus (Maybe Bool)

The domain deletion status. True if a delete request has been received for the domain but resource cleanup is still in progress. False if the domain has not been deleted. Once domain deletion is complete, the status of the domain is no longer returned.

elasticsearchDomainStatus_domainId :: Lens' ElasticsearchDomainStatus Text

The unique identifier for the specified Elasticsearch domain.

elasticsearchDomainStatus_domainName :: Lens' ElasticsearchDomainStatus Text

The name of an Elasticsearch domain. Domain names are unique across the domains owned by an account within an AWS region. Domain names start with a letter or number and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

elasticsearchDomainStatus_arn :: Lens' ElasticsearchDomainStatus Text

The Amazon resource name (ARN) of an Elasticsearch domain. See Identifiers for IAM Entities in Using AWS Identity and Access Management for more information.

elasticsearchDomainStatus_elasticsearchClusterConfig :: Lens' ElasticsearchDomainStatus ElasticsearchClusterConfig

The type and number of instances in the domain cluster.