Network.AWS.ElasticSearch.Types.OptionStatus

Description

Documentation

data OptionStatus

Provides the current status of the entity.

See: newOptionStatus smart constructor.

Constructors

OptionStatus'

Fields

updateVersion :: Maybe Natural

Specifies the latest version for the entity.

pendingDeletion :: Maybe Bool

Indicates whether the Elasticsearch domain is being deleted.

creationDate :: POSIX

Timestamp which tells the creation date for the entity.

updateDate :: POSIX

Timestamp which tells the last updated time for the entity.

state :: OptionState

Provides the OptionState for the Elasticsearch domain.

newOptionStatus

Arguments

:: UTCTime
$sel:creationDate:OptionStatus'

-> UTCTime
$sel:updateDate:OptionStatus'

-> OptionState
$sel:state:OptionStatus'

-> OptionStatus

Create a value of OptionStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:updateVersion:OptionStatus', optionStatus_updateVersion - Specifies the latest version for the entity.

$sel:pendingDeletion:OptionStatus', optionStatus_pendingDeletion - Indicates whether the Elasticsearch domain is being deleted.

$sel:creationDate:OptionStatus', optionStatus_creationDate - Timestamp which tells the creation date for the entity.

$sel:updateDate:OptionStatus', optionStatus_updateDate - Timestamp which tells the last updated time for the entity.

$sel:state:OptionStatus', optionStatus_state - Provides the OptionState for the Elasticsearch domain.

optionStatus_updateVersion :: Lens' OptionStatus (Maybe Natural)

Specifies the latest version for the entity.

optionStatus_pendingDeletion :: Lens' OptionStatus (Maybe Bool)

Indicates whether the Elasticsearch domain is being deleted.

optionStatus_creationDate :: Lens' OptionStatus UTCTime

Timestamp which tells the creation date for the entity.

optionStatus_updateDate :: Lens' OptionStatus UTCTime

Timestamp which tells the last updated time for the entity.

optionStatus_state :: Lens' OptionStatus OptionState

Provides the OptionState for the Elasticsearch domain.