Network.AWS.ElasticSearch.Types.OutboundCrossClusterSearchConnectionStatus

Description

Documentation

data OutboundCrossClusterSearchConnectionStatus

Specifies the connection status of an outbound cross-cluster search connection.

See: newOutboundCrossClusterSearchConnectionStatus smart constructor.

Constructors

OutboundCrossClusterSearchConnectionStatus'

Fields

message :: Maybe Text

Specifies verbose information for the outbound connection status.

statusCode :: Maybe OutboundCrossClusterSearchConnectionStatusCode

The state code for outbound connection. This can be one of the following:

VALIDATING: The outbound connection request is being validated.

VALIDATION_FAILED: Validation failed for the connection request.

PENDING_ACCEPTANCE: Outbound connection request is validated and is not yet accepted by destination domain owner.

PROVISIONING: Outbound connection request is in process.

ACTIVE: Outbound connection is active and ready to use.

REJECTED: Outbound connection request is rejected by destination domain owner.

DELETING: Outbound connection deletion is in progress.

DELETED: Outbound connection is deleted and cannot be used further.

newOutboundCrossClusterSearchConnectionStatus :: OutboundCrossClusterSearchConnectionStatus

Create a value of OutboundCrossClusterSearchConnectionStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:message:OutboundCrossClusterSearchConnectionStatus', outboundCrossClusterSearchConnectionStatus_message - Specifies verbose information for the outbound connection status.

$sel:statusCode:OutboundCrossClusterSearchConnectionStatus', outboundCrossClusterSearchConnectionStatus_statusCode - The state code for outbound connection. This can be one of the following:

VALIDATING: The outbound connection request is being validated.

VALIDATION_FAILED: Validation failed for the connection request.

PENDING_ACCEPTANCE: Outbound connection request is validated and is not yet accepted by destination domain owner.

PROVISIONING: Outbound connection request is in process.

ACTIVE: Outbound connection is active and ready to use.

REJECTED: Outbound connection request is rejected by destination domain owner.

DELETING: Outbound connection deletion is in progress.

DELETED: Outbound connection is deleted and cannot be used further.

outboundCrossClusterSearchConnectionStatus_message :: Lens' OutboundCrossClusterSearchConnectionStatus (Maybe Text)

Specifies verbose information for the outbound connection status.

outboundCrossClusterSearchConnectionStatus_statusCode :: Lens' OutboundCrossClusterSearchConnectionStatus (Maybe OutboundCrossClusterSearchConnectionStatusCode)

The state code for outbound connection. This can be one of the following:

VALIDATING: The outbound connection request is being validated.

VALIDATION_FAILED: Validation failed for the connection request.

PENDING_ACCEPTANCE: Outbound connection request is validated and is not yet accepted by destination domain owner.

PROVISIONING: Outbound connection request is in process.

ACTIVE: Outbound connection is active and ready to use.

REJECTED: Outbound connection request is rejected by destination domain owner.

DELETING: Outbound connection deletion is in progress.

DELETED: Outbound connection is deleted and cannot be used further.