Network.AWS.ElasticSearch.Types.PackageDetails

Description

Documentation

data PackageDetails

Basic information about a package.

See: newPackageDetails smart constructor.

Constructors

PackageDetails'

Fields

availablePackageVersion :: Maybe Text

packageStatus :: Maybe PackageStatus

Current state of the package. Values are COPYING/COPY_FAILED/AVAILABLE/DELETING/DELETE_FAILED

createdAt :: Maybe POSIX

Timestamp which tells creation date of the package.

packageName :: Maybe Text

User specified name of the package.

packageID :: Maybe Text

Internal ID of the package.

packageDescription :: Maybe Text

User-specified description of the package.

lastUpdatedAt :: Maybe POSIX

packageType :: Maybe PackageType

Currently supports only TXT-DICTIONARY.

errorDetails :: Maybe ErrorDetails

Additional information if the package is in an error state. Null otherwise.

newPackageDetails :: PackageDetails

Create a value of PackageDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:availablePackageVersion:PackageDetails', packageDetails_availablePackageVersion - Undocumented member.

$sel:packageStatus:PackageDetails', packageDetails_packageStatus - Current state of the package. Values are COPYING/COPY_FAILED/AVAILABLE/DELETING/DELETE_FAILED

$sel:createdAt:PackageDetails', packageDetails_createdAt - Timestamp which tells creation date of the package.

$sel:packageName:PackageDetails', packageDetails_packageName - User specified name of the package.

$sel:packageID:PackageDetails', packageDetails_packageID - Internal ID of the package.

$sel:packageDescription:PackageDetails', packageDetails_packageDescription - User-specified description of the package.

$sel:lastUpdatedAt:PackageDetails', packageDetails_lastUpdatedAt - Undocumented member.

$sel:packageType:PackageDetails', packageDetails_packageType - Currently supports only TXT-DICTIONARY.

$sel:errorDetails:PackageDetails', packageDetails_errorDetails - Additional information if the package is in an error state. Null otherwise.

packageDetails_availablePackageVersion :: Lens' PackageDetails (Maybe Text)

Undocumented member.

packageDetails_packageStatus :: Lens' PackageDetails (Maybe PackageStatus)

Current state of the package. Values are COPYING/COPY_FAILED/AVAILABLE/DELETING/DELETE_FAILED

packageDetails_createdAt :: Lens' PackageDetails (Maybe UTCTime)

Timestamp which tells creation date of the package.

packageDetails_packageName :: Lens' PackageDetails (Maybe Text)

User specified name of the package.

packageDetails_packageID :: Lens' PackageDetails (Maybe Text)

Internal ID of the package.

packageDetails_packageDescription :: Lens' PackageDetails (Maybe Text)

User-specified description of the package.

packageDetails_lastUpdatedAt :: Lens' PackageDetails (Maybe UTCTime)

Undocumented member.

packageDetails_packageType :: Lens' PackageDetails (Maybe PackageType)

Currently supports only TXT-DICTIONARY.

packageDetails_errorDetails :: Lens' PackageDetails (Maybe ErrorDetails)

Additional information if the package is in an error state. Null otherwise.