Network.AWS.ElasticSearch.Types.ReservedElasticsearchInstance

Description

Documentation

data ReservedElasticsearchInstance

Details of a reserved Elasticsearch instance.

See: newReservedElasticsearchInstance smart constructor.

Constructors

ReservedElasticsearchInstance'

Fields

reservationName :: Maybe Text

The customer-specified identifier to track this reservation.

paymentOption :: Maybe ReservedElasticsearchInstancePaymentOption

The payment option as defined in the reserved Elasticsearch instance offering.

elasticsearchInstanceCount :: Maybe Int

The number of Elasticsearch instances that have been reserved.

duration :: Maybe Int

The duration, in seconds, for which the Elasticsearch instance is reserved.

startTime :: Maybe POSIX

The time the reservation started.

currencyCode :: Maybe Text

The currency code for the reserved Elasticsearch instance offering.

elasticsearchInstanceType :: Maybe ESPartitionInstanceType

The Elasticsearch instance type offered by the reserved instance offering.

state :: Maybe Text

The state of the reserved Elasticsearch instance.

fixedPrice :: Maybe Double

The upfront fixed charge you will paid to purchase the specific reserved Elasticsearch instance offering.

reservedElasticsearchInstanceId :: Maybe Text

The unique identifier for the reservation.

reservedElasticsearchInstanceOfferingId :: Maybe Text

The offering identifier.

usagePrice :: Maybe Double

The rate you are charged for each hour for the domain that is using this reserved instance.

recurringCharges :: Maybe [RecurringCharge]

The charge to your account regardless of whether you are creating any domains using the instance offering.

newReservedElasticsearchInstance :: ReservedElasticsearchInstance

Create a value of ReservedElasticsearchInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reservationName:ReservedElasticsearchInstance', reservedElasticsearchInstance_reservationName - The customer-specified identifier to track this reservation.

$sel:paymentOption:ReservedElasticsearchInstance', reservedElasticsearchInstance_paymentOption - The payment option as defined in the reserved Elasticsearch instance offering.

$sel:elasticsearchInstanceCount:ReservedElasticsearchInstance', reservedElasticsearchInstance_elasticsearchInstanceCount - The number of Elasticsearch instances that have been reserved.

$sel:duration:ReservedElasticsearchInstance', reservedElasticsearchInstance_duration - The duration, in seconds, for which the Elasticsearch instance is reserved.

$sel:startTime:ReservedElasticsearchInstance', reservedElasticsearchInstance_startTime - The time the reservation started.

$sel:currencyCode:ReservedElasticsearchInstance', reservedElasticsearchInstance_currencyCode - The currency code for the reserved Elasticsearch instance offering.

$sel:elasticsearchInstanceType:ReservedElasticsearchInstance', reservedElasticsearchInstance_elasticsearchInstanceType - The Elasticsearch instance type offered by the reserved instance offering.

$sel:state:ReservedElasticsearchInstance', reservedElasticsearchInstance_state - The state of the reserved Elasticsearch instance.

$sel:fixedPrice:ReservedElasticsearchInstance', reservedElasticsearchInstance_fixedPrice - The upfront fixed charge you will paid to purchase the specific reserved Elasticsearch instance offering.

$sel:reservedElasticsearchInstanceId:ReservedElasticsearchInstance', reservedElasticsearchInstance_reservedElasticsearchInstanceId - The unique identifier for the reservation.

$sel:reservedElasticsearchInstanceOfferingId:ReservedElasticsearchInstance', reservedElasticsearchInstance_reservedElasticsearchInstanceOfferingId - The offering identifier.

$sel:usagePrice:ReservedElasticsearchInstance', reservedElasticsearchInstance_usagePrice - The rate you are charged for each hour for the domain that is using this reserved instance.

$sel:recurringCharges:ReservedElasticsearchInstance', reservedElasticsearchInstance_recurringCharges - The charge to your account regardless of whether you are creating any domains using the instance offering.

reservedElasticsearchInstance_reservationName :: Lens' ReservedElasticsearchInstance (Maybe Text)

The customer-specified identifier to track this reservation.

reservedElasticsearchInstance_paymentOption :: Lens' ReservedElasticsearchInstance (Maybe ReservedElasticsearchInstancePaymentOption)

The payment option as defined in the reserved Elasticsearch instance offering.

reservedElasticsearchInstance_elasticsearchInstanceCount :: Lens' ReservedElasticsearchInstance (Maybe Int)

The number of Elasticsearch instances that have been reserved.

reservedElasticsearchInstance_duration :: Lens' ReservedElasticsearchInstance (Maybe Int)

The duration, in seconds, for which the Elasticsearch instance is reserved.

reservedElasticsearchInstance_startTime :: Lens' ReservedElasticsearchInstance (Maybe UTCTime)

The time the reservation started.

reservedElasticsearchInstance_currencyCode :: Lens' ReservedElasticsearchInstance (Maybe Text)

The currency code for the reserved Elasticsearch instance offering.

reservedElasticsearchInstance_elasticsearchInstanceType :: Lens' ReservedElasticsearchInstance (Maybe ESPartitionInstanceType)

The Elasticsearch instance type offered by the reserved instance offering.

reservedElasticsearchInstance_state :: Lens' ReservedElasticsearchInstance (Maybe Text)

The state of the reserved Elasticsearch instance.

reservedElasticsearchInstance_fixedPrice :: Lens' ReservedElasticsearchInstance (Maybe Double)

The upfront fixed charge you will paid to purchase the specific reserved Elasticsearch instance offering.

reservedElasticsearchInstance_reservedElasticsearchInstanceId :: Lens' ReservedElasticsearchInstance (Maybe Text)

The unique identifier for the reservation.

reservedElasticsearchInstance_reservedElasticsearchInstanceOfferingId :: Lens' ReservedElasticsearchInstance (Maybe Text)

The offering identifier.

reservedElasticsearchInstance_usagePrice :: Lens' ReservedElasticsearchInstance (Maybe Double)

The rate you are charged for each hour for the domain that is using this reserved instance.

reservedElasticsearchInstance_recurringCharges :: Lens' ReservedElasticsearchInstance (Maybe [RecurringCharge])

The charge to your account regardless of whether you are creating any domains using the instance offering.