Network.AWS.ElasticSearch.Types.ReservedElasticsearchInstanceOffering

Description

Documentation

data ReservedElasticsearchInstanceOffering

Details of a reserved Elasticsearch instance offering.

See: newReservedElasticsearchInstanceOffering smart constructor.

Constructors

ReservedElasticsearchInstanceOffering'

Fields

paymentOption :: Maybe ReservedElasticsearchInstancePaymentOption

Payment option for the reserved Elasticsearch instance offering

duration :: Maybe Int

The duration, in seconds, for which the offering will reserve the Elasticsearch instance.

currencyCode :: Maybe Text

The currency code for the reserved Elasticsearch instance offering.

elasticsearchInstanceType :: Maybe ESPartitionInstanceType

The Elasticsearch instance type offered by the reserved instance offering.

fixedPrice :: Maybe Double

The upfront fixed charge you will pay to purchase the specific reserved Elasticsearch instance offering.

reservedElasticsearchInstanceOfferingId :: Maybe Text

The Elasticsearch reserved instance offering identifier.

usagePrice :: Maybe Double

The rate you are charged for each hour the domain that is using the offering is running.

recurringCharges :: Maybe [RecurringCharge]

The charge to your account regardless of whether you are creating any domains using the instance offering.

newReservedElasticsearchInstanceOffering :: ReservedElasticsearchInstanceOffering

Create a value of ReservedElasticsearchInstanceOffering with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:paymentOption:ReservedElasticsearchInstanceOffering', reservedElasticsearchInstanceOffering_paymentOption - Payment option for the reserved Elasticsearch instance offering

$sel:duration:ReservedElasticsearchInstanceOffering', reservedElasticsearchInstanceOffering_duration - The duration, in seconds, for which the offering will reserve the Elasticsearch instance.

$sel:currencyCode:ReservedElasticsearchInstanceOffering', reservedElasticsearchInstanceOffering_currencyCode - The currency code for the reserved Elasticsearch instance offering.

$sel:elasticsearchInstanceType:ReservedElasticsearchInstanceOffering', reservedElasticsearchInstanceOffering_elasticsearchInstanceType - The Elasticsearch instance type offered by the reserved instance offering.

$sel:fixedPrice:ReservedElasticsearchInstanceOffering', reservedElasticsearchInstanceOffering_fixedPrice - The upfront fixed charge you will pay to purchase the specific reserved Elasticsearch instance offering.

$sel:reservedElasticsearchInstanceOfferingId:ReservedElasticsearchInstanceOffering', reservedElasticsearchInstanceOffering_reservedElasticsearchInstanceOfferingId - The Elasticsearch reserved instance offering identifier.

$sel:usagePrice:ReservedElasticsearchInstanceOffering', reservedElasticsearchInstanceOffering_usagePrice - The rate you are charged for each hour the domain that is using the offering is running.

$sel:recurringCharges:ReservedElasticsearchInstanceOffering', reservedElasticsearchInstanceOffering_recurringCharges - The charge to your account regardless of whether you are creating any domains using the instance offering.

reservedElasticsearchInstanceOffering_paymentOption :: Lens' ReservedElasticsearchInstanceOffering (Maybe ReservedElasticsearchInstancePaymentOption)

Payment option for the reserved Elasticsearch instance offering

reservedElasticsearchInstanceOffering_duration :: Lens' ReservedElasticsearchInstanceOffering (Maybe Int)

The duration, in seconds, for which the offering will reserve the Elasticsearch instance.

reservedElasticsearchInstanceOffering_currencyCode :: Lens' ReservedElasticsearchInstanceOffering (Maybe Text)

The currency code for the reserved Elasticsearch instance offering.

reservedElasticsearchInstanceOffering_elasticsearchInstanceType :: Lens' ReservedElasticsearchInstanceOffering (Maybe ESPartitionInstanceType)

The Elasticsearch instance type offered by the reserved instance offering.

reservedElasticsearchInstanceOffering_fixedPrice :: Lens' ReservedElasticsearchInstanceOffering (Maybe Double)

The upfront fixed charge you will pay to purchase the specific reserved Elasticsearch instance offering.

reservedElasticsearchInstanceOffering_reservedElasticsearchInstanceOfferingId :: Lens' ReservedElasticsearchInstanceOffering (Maybe Text)

The Elasticsearch reserved instance offering identifier.

reservedElasticsearchInstanceOffering_usagePrice :: Lens' ReservedElasticsearchInstanceOffering (Maybe Double)

The rate you are charged for each hour the domain that is using the offering is running.

reservedElasticsearchInstanceOffering_recurringCharges :: Lens' ReservedElasticsearchInstanceOffering (Maybe [RecurringCharge])

The charge to your account regardless of whether you are creating any domains using the instance offering.