Network.AWS.ElasticSearch.UpgradeElasticsearchDomain

Description

Allows you to either upgrade your domain or perform an Upgrade eligibility check to a compatible Elasticsearch version.

Creating a Request

data UpgradeElasticsearchDomain

Container for request parameters to UpgradeElasticsearchDomain operation.

See: newUpgradeElasticsearchDomain smart constructor.

Constructors

UpgradeElasticsearchDomain'

Fields

performCheckOnly :: Maybe Bool

This flag, when set to True, indicates that an Upgrade Eligibility Check needs to be performed. This will not actually perform the Upgrade.

domainName :: Text

targetVersion :: Text

The version of Elasticsearch that you intend to upgrade the domain to.

newUpgradeElasticsearchDomain

Arguments

:: Text
$sel:domainName:UpgradeElasticsearchDomain'

-> Text
$sel:targetVersion:UpgradeElasticsearchDomain'

-> UpgradeElasticsearchDomain

Create a value of UpgradeElasticsearchDomain with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:performCheckOnly:UpgradeElasticsearchDomain', upgradeElasticsearchDomain_performCheckOnly - This flag, when set to True, indicates that an Upgrade Eligibility Check needs to be performed. This will not actually perform the Upgrade.

$sel:domainName:UpgradeElasticsearchDomain', upgradeElasticsearchDomain_domainName - Undocumented member.

$sel:targetVersion:UpgradeElasticsearchDomain', upgradeElasticsearchDomain_targetVersion - The version of Elasticsearch that you intend to upgrade the domain to.

Request Lenses

upgradeElasticsearchDomain_performCheckOnly :: Lens' UpgradeElasticsearchDomain (Maybe Bool)

This flag, when set to True, indicates that an Upgrade Eligibility Check needs to be performed. This will not actually perform the Upgrade.

upgradeElasticsearchDomain_domainName :: Lens' UpgradeElasticsearchDomain Text

Undocumented member.

upgradeElasticsearchDomain_targetVersion :: Lens' UpgradeElasticsearchDomain Text

The version of Elasticsearch that you intend to upgrade the domain to.

Destructuring the Response

data UpgradeElasticsearchDomainResponse

Container for response returned by UpgradeElasticsearchDomain operation.

See: newUpgradeElasticsearchDomainResponse smart constructor.

Constructors

UpgradeElasticsearchDomainResponse'

Fields

domainName :: Maybe Text

performCheckOnly :: Maybe Bool

This flag, when set to True, indicates that an Upgrade Eligibility Check needs to be performed. This will not actually perform the Upgrade.

targetVersion :: Maybe Text

The version of Elasticsearch that you intend to upgrade the domain to.

httpStatus :: Int

The response's http status code.

newUpgradeElasticsearchDomainResponse

Arguments

:: Int
$sel:httpStatus:UpgradeElasticsearchDomainResponse'

-> UpgradeElasticsearchDomainResponse

Create a value of UpgradeElasticsearchDomainResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainName:UpgradeElasticsearchDomain', upgradeElasticsearchDomainResponse_domainName - Undocumented member.

$sel:performCheckOnly:UpgradeElasticsearchDomain', upgradeElasticsearchDomainResponse_performCheckOnly - This flag, when set to True, indicates that an Upgrade Eligibility Check needs to be performed. This will not actually perform the Upgrade.

$sel:targetVersion:UpgradeElasticsearchDomain', upgradeElasticsearchDomainResponse_targetVersion - The version of Elasticsearch that you intend to upgrade the domain to.

$sel:httpStatus:UpgradeElasticsearchDomainResponse', upgradeElasticsearchDomainResponse_httpStatus - The response's http status code.

Response Lenses

upgradeElasticsearchDomainResponse_domainName :: Lens' UpgradeElasticsearchDomainResponse (Maybe Text)

Undocumented member.

upgradeElasticsearchDomainResponse_performCheckOnly :: Lens' UpgradeElasticsearchDomainResponse (Maybe Bool)

This flag, when set to True, indicates that an Upgrade Eligibility Check needs to be performed. This will not actually perform the Upgrade.

upgradeElasticsearchDomainResponse_targetVersion :: Lens' UpgradeElasticsearchDomainResponse (Maybe Text)

The version of Elasticsearch that you intend to upgrade the domain to.

upgradeElasticsearchDomainResponse_httpStatus :: Lens' UpgradeElasticsearchDomainResponse Int

The response's http status code.