Network.AWS.ElasticTranscoder.CreateJob

Description

When you create a job, Elastic Transcoder returns JSON data that includes the values that you specified plus information about the job that is created.

If you have specified more than one output for your jobs (for example, one output for the Kindle Fire and another output for the Apple iPhone 4s), you currently must use the Elastic Transcoder API to list the jobs (as opposed to the AWS Console).

Creating a Request

data CreateJob

The CreateJobRequest structure.

See: newCreateJob smart constructor.

Constructors

CreateJob'

Fields

inputs :: Maybe [JobInput]

A section of the request body that provides information about the files that are being transcoded.

input :: Maybe JobInput

A section of the request body that provides information about the file that is being transcoded.

userMetadata :: Maybe (HashMap Text Text)

User-defined metadata that you want to associate with an Elastic Transcoder job. You specify metadata in key/value pairs, and you can add up to 10 key/value pairs per job. Elastic Transcoder does not guarantee that key/value pairs are returned in the same order in which you specify them.

outputs :: Maybe [CreateJobOutput]

A section of the request body that provides information about the transcoded (target) files. We recommend that you use the Outputs syntax instead of the Output syntax.

output :: Maybe CreateJobOutput

A section of the request body that provides information about the transcoded (target) file. We strongly recommend that you use the Outputs syntax instead of the Output syntax.

playlists :: Maybe [CreateJobPlaylist]

If you specify a preset in PresetId for which the value of Container is fmp4 (Fragmented MP4) or ts (MPEG-TS), Playlists contains information about the master playlists that you want Elastic Transcoder to create.

The maximum number of master playlists in a job is 30.

outputKeyPrefix :: Maybe Text

The value, if any, that you want Elastic Transcoder to prepend to the names of all files that this job creates, including output files, thumbnails, and playlists.

pipelineId :: Text

The Id of the pipeline that you want Elastic Transcoder to use for transcoding. The pipeline determines several settings, including the Amazon S3 bucket from which Elastic Transcoder gets the files to transcode and the bucket into which Elastic Transcoder puts the transcoded files.

newCreateJob

Arguments

:: Text
$sel:pipelineId:CreateJob'

-> CreateJob

Create a value of CreateJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:inputs:CreateJob', createJob_inputs - A section of the request body that provides information about the files that are being transcoded.

$sel:input:CreateJob', createJob_input - A section of the request body that provides information about the file that is being transcoded.

$sel:userMetadata:CreateJob', createJob_userMetadata - User-defined metadata that you want to associate with an Elastic Transcoder job. You specify metadata in key/value pairs, and you can add up to 10 key/value pairs per job. Elastic Transcoder does not guarantee that key/value pairs are returned in the same order in which you specify them.

$sel:outputs:CreateJob', createJob_outputs - A section of the request body that provides information about the transcoded (target) files. We recommend that you use the Outputs syntax instead of the Output syntax.

$sel:output:CreateJob', createJob_output - A section of the request body that provides information about the transcoded (target) file. We strongly recommend that you use the Outputs syntax instead of the Output syntax.

$sel:playlists:CreateJob', createJob_playlists - If you specify a preset in PresetId for which the value of Container is fmp4 (Fragmented MP4) or ts (MPEG-TS), Playlists contains information about the master playlists that you want Elastic Transcoder to create.

The maximum number of master playlists in a job is 30.

$sel:outputKeyPrefix:CreateJob', createJob_outputKeyPrefix - The value, if any, that you want Elastic Transcoder to prepend to the names of all files that this job creates, including output files, thumbnails, and playlists.

$sel:pipelineId:CreateJob', createJob_pipelineId - The Id of the pipeline that you want Elastic Transcoder to use for transcoding. The pipeline determines several settings, including the Amazon S3 bucket from which Elastic Transcoder gets the files to transcode and the bucket into which Elastic Transcoder puts the transcoded files.

Request Lenses

createJob_inputs :: Lens' CreateJob (Maybe [JobInput])

A section of the request body that provides information about the files that are being transcoded.

createJob_input :: Lens' CreateJob (Maybe JobInput)

A section of the request body that provides information about the file that is being transcoded.

createJob_userMetadata :: Lens' CreateJob (Maybe (HashMap Text Text))

User-defined metadata that you want to associate with an Elastic Transcoder job. You specify metadata in key/value pairs, and you can add up to 10 key/value pairs per job. Elastic Transcoder does not guarantee that key/value pairs are returned in the same order in which you specify them.

createJob_outputs :: Lens' CreateJob (Maybe [CreateJobOutput])

A section of the request body that provides information about the transcoded (target) files. We recommend that you use the Outputs syntax instead of the Output syntax.

createJob_output :: Lens' CreateJob (Maybe CreateJobOutput)

A section of the request body that provides information about the transcoded (target) file. We strongly recommend that you use the Outputs syntax instead of the Output syntax.

createJob_playlists :: Lens' CreateJob (Maybe [CreateJobPlaylist])

If you specify a preset in PresetId for which the value of Container is fmp4 (Fragmented MP4) or ts (MPEG-TS), Playlists contains information about the master playlists that you want Elastic Transcoder to create.

The maximum number of master playlists in a job is 30.

createJob_outputKeyPrefix :: Lens' CreateJob (Maybe Text)

The value, if any, that you want Elastic Transcoder to prepend to the names of all files that this job creates, including output files, thumbnails, and playlists.

createJob_pipelineId :: Lens' CreateJob Text

The Id of the pipeline that you want Elastic Transcoder to use for transcoding. The pipeline determines several settings, including the Amazon S3 bucket from which Elastic Transcoder gets the files to transcode and the bucket into which Elastic Transcoder puts the transcoded files.

Destructuring the Response

data CreateJobResponse

The CreateJobResponse structure.

See: newCreateJobResponse smart constructor.

Constructors

CreateJobResponse'

Fields

job :: Maybe Job

A section of the response body that provides information about the job that is created.

httpStatus :: Int

The response's http status code.

newCreateJobResponse

Arguments

:: Int
$sel:httpStatus:CreateJobResponse'

-> CreateJobResponse

Create a value of CreateJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:job:CreateJobResponse', createJobResponse_job - A section of the response body that provides information about the job that is created.

$sel:httpStatus:CreateJobResponse', createJobResponse_httpStatus - The response's http status code.

Response Lenses

createJobResponse_job :: Lens' CreateJobResponse (Maybe Job)

A section of the response body that provides information about the job that is created.

createJobResponse_httpStatus :: Lens' CreateJobResponse Int

The response's http status code.