Network.AWS.ElasticTranscoder.CreatePreset

Description

The CreatePreset operation creates a preset with settings that you specify.

Elastic Transcoder checks the CreatePreset settings to ensure that they meet Elastic Transcoder requirements and to determine whether they comply with H.264 standards. If your settings are not valid for Elastic Transcoder, Elastic Transcoder returns an HTTP 400 response (ValidationException) and does not create the preset. If the settings are valid for Elastic Transcoder but aren't strictly compliant with the H.264 standard, Elastic Transcoder creates the preset and returns a warning message in the response. This helps you determine whether your settings comply with the H.264 standard while giving you greater flexibility with respect to the video that Elastic Transcoder produces.

Elastic Transcoder uses the H.264 video-compression format. For more information, see the International Telecommunication Union publication /Recommendation ITU-T H.264: Advanced video coding for generic audiovisual services/.

Creating a Request

data CreatePreset

The CreatePresetRequest structure.

See: newCreatePreset smart constructor.

Constructors

CreatePreset'

Fields

video :: Maybe VideoParameters

A section of the request body that specifies the video parameters.

thumbnails :: Maybe Thumbnails

A section of the request body that specifies the thumbnail parameters, if any.

description :: Maybe Text

A description of the preset.

audio :: Maybe AudioParameters

A section of the request body that specifies the audio parameters.

name :: Text

The name of the preset. We recommend that the name be unique within the AWS account, but uniqueness is not enforced.

container :: Text

The container type for the output file. Valid values include flac, flv, fmp4, gif, mp3, mp4, mpg, mxf, oga, ogg, ts, and webm.

newCreatePreset

Arguments

:: Text
$sel:name:CreatePreset'

-> Text
$sel:container:CreatePreset'

-> CreatePreset

Create a value of CreatePreset with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:video:CreatePreset', createPreset_video - A section of the request body that specifies the video parameters.

$sel:thumbnails:CreatePreset', createPreset_thumbnails - A section of the request body that specifies the thumbnail parameters, if any.

$sel:description:CreatePreset', createPreset_description - A description of the preset.

$sel:audio:CreatePreset', createPreset_audio - A section of the request body that specifies the audio parameters.

$sel:name:CreatePreset', createPreset_name - The name of the preset. We recommend that the name be unique within the AWS account, but uniqueness is not enforced.

$sel:container:CreatePreset', createPreset_container - The container type for the output file. Valid values include flac, flv, fmp4, gif, mp3, mp4, mpg, mxf, oga, ogg, ts, and webm.

Request Lenses

createPreset_video :: Lens' CreatePreset (Maybe VideoParameters)

A section of the request body that specifies the video parameters.

createPreset_thumbnails :: Lens' CreatePreset (Maybe Thumbnails)

A section of the request body that specifies the thumbnail parameters, if any.

createPreset_description :: Lens' CreatePreset (Maybe Text)

A description of the preset.

createPreset_audio :: Lens' CreatePreset (Maybe AudioParameters)

A section of the request body that specifies the audio parameters.

createPreset_name :: Lens' CreatePreset Text

The name of the preset. We recommend that the name be unique within the AWS account, but uniqueness is not enforced.

createPreset_container :: Lens' CreatePreset Text

The container type for the output file. Valid values include flac, flv, fmp4, gif, mp3, mp4, mpg, mxf, oga, ogg, ts, and webm.

Destructuring the Response

data CreatePresetResponse

The CreatePresetResponse structure.

See: newCreatePresetResponse smart constructor.

Constructors

CreatePresetResponse'

Fields

warning :: Maybe Text

If the preset settings don't comply with the standards for the video codec but Elastic Transcoder created the preset, this message explains the reason the preset settings don't meet the standard. Elastic Transcoder created the preset because the settings might produce acceptable output.

preset :: Maybe Preset

A section of the response body that provides information about the preset that is created.

httpStatus :: Int

The response's http status code.

newCreatePresetResponse

Arguments

:: Int
$sel:httpStatus:CreatePresetResponse'

-> CreatePresetResponse

Create a value of CreatePresetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:warning:CreatePresetResponse', createPresetResponse_warning - If the preset settings don't comply with the standards for the video codec but Elastic Transcoder created the preset, this message explains the reason the preset settings don't meet the standard. Elastic Transcoder created the preset because the settings might produce acceptable output.

$sel:preset:CreatePresetResponse', createPresetResponse_preset - A section of the response body that provides information about the preset that is created.

$sel:httpStatus:CreatePresetResponse', createPresetResponse_httpStatus - The response's http status code.

Response Lenses

createPresetResponse_warning :: Lens' CreatePresetResponse (Maybe Text)

If the preset settings don't comply with the standards for the video codec but Elastic Transcoder created the preset, this message explains the reason the preset settings don't meet the standard. Elastic Transcoder created the preset because the settings might produce acceptable output.

createPresetResponse_preset :: Lens' CreatePresetResponse (Maybe Preset)

A section of the response body that provides information about the preset that is created.

createPresetResponse_httpStatus :: Lens' CreatePresetResponse Int

The response's http status code.