Network.AWS.ElasticTranscoder.ListJobsByStatus

Description

The ListJobsByStatus operation gets a list of jobs that have a specified status. The response body contains one element for each job that satisfies the search criteria.

This operation returns paginated results.

Creating a Request

data ListJobsByStatus

The ListJobsByStatusRequest structure.

See: newListJobsByStatus smart constructor.

Constructors

ListJobsByStatus'

Fields

ascending :: Maybe Text

To list jobs in chronological order by the date and time that they were submitted, enter true. To list jobs in reverse chronological order, enter false.

pageToken :: Maybe Text

When Elastic Transcoder returns more than one page of results, use pageToken in subsequent GET requests to get each successive page of results.

status :: Text

To get information about all of the jobs associated with the current AWS account that have a given status, specify the following status: Submitted, Progressing, Complete, Canceled, or Error.

newListJobsByStatus

Arguments

:: Text
$sel:status:ListJobsByStatus'

-> ListJobsByStatus

Create a value of ListJobsByStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ascending:ListJobsByStatus', listJobsByStatus_ascending - To list jobs in chronological order by the date and time that they were submitted, enter true. To list jobs in reverse chronological order, enter false.

$sel:pageToken:ListJobsByStatus', listJobsByStatus_pageToken - When Elastic Transcoder returns more than one page of results, use pageToken in subsequent GET requests to get each successive page of results.

$sel:status:ListJobsByStatus', listJobsByStatus_status - To get information about all of the jobs associated with the current AWS account that have a given status, specify the following status: Submitted, Progressing, Complete, Canceled, or Error.

Request Lenses

listJobsByStatus_ascending :: Lens' ListJobsByStatus (Maybe Text)

To list jobs in chronological order by the date and time that they were submitted, enter true. To list jobs in reverse chronological order, enter false.

listJobsByStatus_pageToken :: Lens' ListJobsByStatus (Maybe Text)

When Elastic Transcoder returns more than one page of results, use pageToken in subsequent GET requests to get each successive page of results.

listJobsByStatus_status :: Lens' ListJobsByStatus Text

To get information about all of the jobs associated with the current AWS account that have a given status, specify the following status: Submitted, Progressing, Complete, Canceled, or Error.

Destructuring the Response

data ListJobsByStatusResponse

The ListJobsByStatusResponse structure.

See: newListJobsByStatusResponse smart constructor.

Constructors

ListJobsByStatusResponse'

Fields

nextPageToken :: Maybe Text

A value that you use to access the second and subsequent pages of results, if any. When the jobs in the specified pipeline fit on one page or when you've reached the last page of results, the value of NextPageToken is null.

jobs :: Maybe [Job]

An array of Job objects that have the specified status.

httpStatus :: Int

The response's http status code.

newListJobsByStatusResponse

Arguments

:: Int
$sel:httpStatus:ListJobsByStatusResponse'

-> ListJobsByStatusResponse

Create a value of ListJobsByStatusResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextPageToken:ListJobsByStatusResponse', listJobsByStatusResponse_nextPageToken - A value that you use to access the second and subsequent pages of results, if any. When the jobs in the specified pipeline fit on one page or when you've reached the last page of results, the value of NextPageToken is null.

$sel:jobs:ListJobsByStatusResponse', listJobsByStatusResponse_jobs - An array of Job objects that have the specified status.

$sel:httpStatus:ListJobsByStatusResponse', listJobsByStatusResponse_httpStatus - The response's http status code.

Response Lenses

listJobsByStatusResponse_nextPageToken :: Lens' ListJobsByStatusResponse (Maybe Text)

A value that you use to access the second and subsequent pages of results, if any. When the jobs in the specified pipeline fit on one page or when you've reached the last page of results, the value of NextPageToken is null.

listJobsByStatusResponse_jobs :: Lens' ListJobsByStatusResponse (Maybe [Job])

An array of Job objects that have the specified status.

listJobsByStatusResponse_httpStatus :: Lens' ListJobsByStatusResponse Int

The response's http status code.