Network.AWS.ElasticTranscoder.ListPipelines

Description

The ListPipelines operation gets a list of the pipelines associated with the current AWS account.

This operation returns paginated results.

Creating a Request

data ListPipelines

The ListPipelineRequest structure.

See: newListPipelines smart constructor.

Constructors

ListPipelines'

Fields

ascending :: Maybe Text

To list pipelines in chronological order by the date and time that they were created, enter true. To list pipelines in reverse chronological order, enter false.

pageToken :: Maybe Text

When Elastic Transcoder returns more than one page of results, use pageToken in subsequent GET requests to get each successive page of results.

newListPipelines :: ListPipelines

Create a value of ListPipelines with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ascending:ListPipelines', listPipelines_ascending - To list pipelines in chronological order by the date and time that they were created, enter true. To list pipelines in reverse chronological order, enter false.

$sel:pageToken:ListPipelines', listPipelines_pageToken - When Elastic Transcoder returns more than one page of results, use pageToken in subsequent GET requests to get each successive page of results.

Request Lenses

listPipelines_ascending :: Lens' ListPipelines (Maybe Text)

To list pipelines in chronological order by the date and time that they were created, enter true. To list pipelines in reverse chronological order, enter false.

listPipelines_pageToken :: Lens' ListPipelines (Maybe Text)

When Elastic Transcoder returns more than one page of results, use pageToken in subsequent GET requests to get each successive page of results.

Destructuring the Response

data ListPipelinesResponse

A list of the pipelines associated with the current AWS account.

See: newListPipelinesResponse smart constructor.

Constructors

ListPipelinesResponse'

Fields

nextPageToken :: Maybe Text

A value that you use to access the second and subsequent pages of results, if any. When the pipelines fit on one page or when you've reached the last page of results, the value of NextPageToken is null.

pipelines :: Maybe [Pipeline]

An array of Pipeline objects.

httpStatus :: Int

The response's http status code.

newListPipelinesResponse

Arguments

:: Int
$sel:httpStatus:ListPipelinesResponse'

-> ListPipelinesResponse

Create a value of ListPipelinesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextPageToken:ListPipelinesResponse', listPipelinesResponse_nextPageToken - A value that you use to access the second and subsequent pages of results, if any. When the pipelines fit on one page or when you've reached the last page of results, the value of NextPageToken is null.

$sel:pipelines:ListPipelinesResponse', listPipelinesResponse_pipelines - An array of Pipeline objects.

$sel:httpStatus:ListPipelinesResponse', listPipelinesResponse_httpStatus - The response's http status code.

Response Lenses

listPipelinesResponse_nextPageToken :: Lens' ListPipelinesResponse (Maybe Text)

A value that you use to access the second and subsequent pages of results, if any. When the pipelines fit on one page or when you've reached the last page of results, the value of NextPageToken is null.

listPipelinesResponse_pipelines :: Lens' ListPipelinesResponse (Maybe [Pipeline])

An array of Pipeline objects.

listPipelinesResponse_httpStatus :: Lens' ListPipelinesResponse Int

The response's http status code.