Network.AWS.ElasticTranscoder.ReadPipeline

Description

The ReadPipeline operation gets detailed information about a pipeline.

Creating a Request

data ReadPipeline

The ReadPipelineRequest structure.

See: newReadPipeline smart constructor.

Constructors

ReadPipeline'

Fields

id :: Text

The identifier of the pipeline to read.

newReadPipeline

Arguments

:: Text
$sel:id:ReadPipeline'

-> ReadPipeline

Create a value of ReadPipeline with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:ReadPipeline', readPipeline_id - The identifier of the pipeline to read.

Request Lenses

readPipeline_id :: Lens' ReadPipeline Text

The identifier of the pipeline to read.

Destructuring the Response

data ReadPipelineResponse

The ReadPipelineResponse structure.

See: newReadPipelineResponse smart constructor.

Constructors

ReadPipelineResponse'

Fields

warnings :: Maybe [Warning]

Elastic Transcoder returns a warning if the resources used by your pipeline are not in the same region as the pipeline.

Using resources in the same region, such as your Amazon S3 buckets, Amazon SNS notification topics, and AWS KMS key, reduces processing time and prevents cross-regional charges.

pipeline :: Maybe Pipeline

A section of the response body that provides information about the pipeline.

httpStatus :: Int

The response's http status code.

newReadPipelineResponse

Arguments

:: Int
$sel:httpStatus:ReadPipelineResponse'

-> ReadPipelineResponse

Create a value of ReadPipelineResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:warnings:ReadPipelineResponse', readPipelineResponse_warnings - Elastic Transcoder returns a warning if the resources used by your pipeline are not in the same region as the pipeline.

Using resources in the same region, such as your Amazon S3 buckets, Amazon SNS notification topics, and AWS KMS key, reduces processing time and prevents cross-regional charges.

$sel:pipeline:ReadPipelineResponse', readPipelineResponse_pipeline - A section of the response body that provides information about the pipeline.

$sel:httpStatus:ReadPipelineResponse', readPipelineResponse_httpStatus - The response's http status code.

Response Lenses

readPipelineResponse_warnings :: Lens' ReadPipelineResponse (Maybe [Warning])

Elastic Transcoder returns a warning if the resources used by your pipeline are not in the same region as the pipeline.

Using resources in the same region, such as your Amazon S3 buckets, Amazon SNS notification topics, and AWS KMS key, reduces processing time and prevents cross-regional charges.

readPipelineResponse_pipeline :: Lens' ReadPipelineResponse (Maybe Pipeline)

A section of the response body that provides information about the pipeline.

readPipelineResponse_httpStatus :: Lens' ReadPipelineResponse Int

The response's http status code.