Network.AWS.ElasticTranscoder.ReadPreset

Description

The ReadPreset operation gets detailed information about a preset.

Creating a Request

data ReadPreset

The ReadPresetRequest structure.

See: newReadPreset smart constructor.

Constructors

ReadPreset'

Fields

id :: Text

The identifier of the preset for which you want to get detailed information.

newReadPreset

Arguments

:: Text
$sel:id:ReadPreset'

-> ReadPreset

Create a value of ReadPreset with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:ReadPreset', readPreset_id - The identifier of the preset for which you want to get detailed information.

Request Lenses

readPreset_id :: Lens' ReadPreset Text

The identifier of the preset for which you want to get detailed information.

Destructuring the Response

data ReadPresetResponse

The ReadPresetResponse structure.

See: newReadPresetResponse smart constructor.

Constructors

ReadPresetResponse'

Fields

preset :: Maybe Preset

A section of the response body that provides information about the preset.

httpStatus :: Int

The response's http status code.

newReadPresetResponse

Arguments

:: Int
$sel:httpStatus:ReadPresetResponse'

-> ReadPresetResponse

Create a value of ReadPresetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:preset:ReadPresetResponse', readPresetResponse_preset - A section of the response body that provides information about the preset.

$sel:httpStatus:ReadPresetResponse', readPresetResponse_httpStatus - The response's http status code.

Response Lenses

readPresetResponse_preset :: Lens' ReadPresetResponse (Maybe Preset)

A section of the response body that provides information about the preset.

readPresetResponse_httpStatus :: Lens' ReadPresetResponse Int

The response's http status code.