Network.AWS.ElasticTranscoder.Types.AudioCodecOptions

Description

Documentation

data AudioCodecOptions

Options associated with your audio codec.

See: newAudioCodecOptions smart constructor.

Constructors

AudioCodecOptions'

Fields

signed :: Maybe Text

You can only choose whether an audio sample is signed when you specify pcm for the value of Audio:Codec.

Whether audio samples are represented with negative and positive numbers (signed) or only positive numbers (unsigned).

The supported value is Signed.

bitDepth :: Maybe Text

You can only choose an audio bit depth when you specify flac or pcm for the value of Audio:Codec.

The bit depth of a sample is how many bits of information are included in the audio samples. The higher the bit depth, the better the audio, but the larger the file.

Valid values are 16 and 24.

The most common bit depth is 24.

profile :: Maybe Text

You can only choose an audio profile when you specify AAC for the value of Audio:Codec.

Specify the AAC profile for the output file. Elastic Transcoder supports the following profiles:

auto: If you specify auto, Elastic Transcoder selects the profile based on the bit rate selected for the output file.

AAC-LC: The most common AAC profile. Use for bit rates larger than 64 kbps.

HE-AAC: Not supported on some older players and devices. Use for bit rates between 40 and 80 kbps.

HE-AACv2: Not supported on some players and devices. Use for bit rates less than 48 kbps.

All outputs in a Smooth playlist must have the same value for Profile.

If you created any presets before AAC profiles were added, Elastic Transcoder automatically updated your presets to use AAC-LC. You can change the value as required.

bitOrder :: Maybe Text

You can only choose an audio bit order when you specify pcm for the value of Audio:Codec.

The order the bits of a PCM sample are stored in.

The supported value is LittleEndian.

newAudioCodecOptions :: AudioCodecOptions

Create a value of AudioCodecOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:signed:AudioCodecOptions', audioCodecOptions_signed - You can only choose whether an audio sample is signed when you specify pcm for the value of Audio:Codec.

Whether audio samples are represented with negative and positive numbers (signed) or only positive numbers (unsigned).

The supported value is Signed.

$sel:bitDepth:AudioCodecOptions', audioCodecOptions_bitDepth - You can only choose an audio bit depth when you specify flac or pcm for the value of Audio:Codec.

The bit depth of a sample is how many bits of information are included in the audio samples. The higher the bit depth, the better the audio, but the larger the file.

Valid values are 16 and 24.

The most common bit depth is 24.

$sel:profile:AudioCodecOptions', audioCodecOptions_profile - You can only choose an audio profile when you specify AAC for the value of Audio:Codec.

Specify the AAC profile for the output file. Elastic Transcoder supports the following profiles:

auto: If you specify auto, Elastic Transcoder selects the profile based on the bit rate selected for the output file.

AAC-LC: The most common AAC profile. Use for bit rates larger than 64 kbps.

HE-AAC: Not supported on some older players and devices. Use for bit rates between 40 and 80 kbps.

HE-AACv2: Not supported on some players and devices. Use for bit rates less than 48 kbps.

All outputs in a Smooth playlist must have the same value for Profile.

If you created any presets before AAC profiles were added, Elastic Transcoder automatically updated your presets to use AAC-LC. You can change the value as required.

$sel:bitOrder:AudioCodecOptions', audioCodecOptions_bitOrder - You can only choose an audio bit order when you specify pcm for the value of Audio:Codec.

The order the bits of a PCM sample are stored in.

The supported value is LittleEndian.

audioCodecOptions_signed :: Lens' AudioCodecOptions (Maybe Text)

You can only choose whether an audio sample is signed when you specify pcm for the value of Audio:Codec.

Whether audio samples are represented with negative and positive numbers (signed) or only positive numbers (unsigned).

The supported value is Signed.

audioCodecOptions_bitDepth :: Lens' AudioCodecOptions (Maybe Text)

You can only choose an audio bit depth when you specify flac or pcm for the value of Audio:Codec.

The bit depth of a sample is how many bits of information are included in the audio samples. The higher the bit depth, the better the audio, but the larger the file.

Valid values are 16 and 24.

The most common bit depth is 24.

audioCodecOptions_profile :: Lens' AudioCodecOptions (Maybe Text)

You can only choose an audio profile when you specify AAC for the value of Audio:Codec.

Specify the AAC profile for the output file. Elastic Transcoder supports the following profiles:

auto: If you specify auto, Elastic Transcoder selects the profile based on the bit rate selected for the output file.

AAC-LC: The most common AAC profile. Use for bit rates larger than 64 kbps.

HE-AAC: Not supported on some older players and devices. Use for bit rates between 40 and 80 kbps.

HE-AACv2: Not supported on some players and devices. Use for bit rates less than 48 kbps.

All outputs in a Smooth playlist must have the same value for Profile.

If you created any presets before AAC profiles were added, Elastic Transcoder automatically updated your presets to use AAC-LC. You can change the value as required.

audioCodecOptions_bitOrder :: Lens' AudioCodecOptions (Maybe Text)

You can only choose an audio bit order when you specify pcm for the value of Audio:Codec.

The order the bits of a PCM sample are stored in.

The supported value is LittleEndian.