Network.AWS.ElasticTranscoder.Types.CaptionSource

Description

Documentation

data CaptionSource

A source file for the input sidecar captions used during the transcoding process.

See: newCaptionSource smart constructor.

Constructors

CaptionSource'

Fields

timeOffset :: Maybe Text

For clip generation or captions that do not start at the same time as the associated video file, the TimeOffset tells Elastic Transcoder how much of the video to encode before including captions.

Specify the TimeOffset in the form [+-]SS.sss or [+-]HH:mm:SS.ss.

encryption :: Maybe Encryption

The encryption settings, if any, that Elastic Transcoder needs to decyrpt your caption sources, or that you want Elastic Transcoder to apply to your caption sources.

key :: Maybe Text

The name of the sidecar caption file that you want Elastic Transcoder to include in the output file.

language :: Maybe Text

A string that specifies the language of the caption. If you specified multiple inputs with captions, the caption language must match in order to be included in the output. Specify this as one of:

2-character ISO 639-1 code

3-character ISO 639-2 code

For more information on ISO language codes and language names, see the List of ISO 639-1 codes.

label :: Maybe Text

The label of the caption shown in the player when choosing a language. We recommend that you put the caption language name here, in the language of the captions.

newCaptionSource :: CaptionSource

Create a value of CaptionSource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timeOffset:CaptionSource', captionSource_timeOffset - For clip generation or captions that do not start at the same time as the associated video file, the TimeOffset tells Elastic Transcoder how much of the video to encode before including captions.

Specify the TimeOffset in the form [+-]SS.sss or [+-]HH:mm:SS.ss.

$sel:encryption:CaptionSource', captionSource_encryption - The encryption settings, if any, that Elastic Transcoder needs to decyrpt your caption sources, or that you want Elastic Transcoder to apply to your caption sources.

$sel:key:CaptionSource', captionSource_key - The name of the sidecar caption file that you want Elastic Transcoder to include in the output file.

$sel:language:CaptionSource', captionSource_language - A string that specifies the language of the caption. If you specified multiple inputs with captions, the caption language must match in order to be included in the output. Specify this as one of:

2-character ISO 639-1 code

3-character ISO 639-2 code

For more information on ISO language codes and language names, see the List of ISO 639-1 codes.

$sel:label:CaptionSource', captionSource_label - The label of the caption shown in the player when choosing a language. We recommend that you put the caption language name here, in the language of the captions.

captionSource_timeOffset :: Lens' CaptionSource (Maybe Text)

For clip generation or captions that do not start at the same time as the associated video file, the TimeOffset tells Elastic Transcoder how much of the video to encode before including captions.

Specify the TimeOffset in the form [+-]SS.sss or [+-]HH:mm:SS.ss.

captionSource_encryption :: Lens' CaptionSource (Maybe Encryption)

The encryption settings, if any, that Elastic Transcoder needs to decyrpt your caption sources, or that you want Elastic Transcoder to apply to your caption sources.

captionSource_key :: Lens' CaptionSource (Maybe Text)

The name of the sidecar caption file that you want Elastic Transcoder to include in the output file.

captionSource_language :: Lens' CaptionSource (Maybe Text)

A string that specifies the language of the caption. If you specified multiple inputs with captions, the caption language must match in order to be included in the output. Specify this as one of:

2-character ISO 639-1 code

3-character ISO 639-2 code

For more information on ISO language codes and language names, see the List of ISO 639-1 codes.

captionSource_label :: Lens' CaptionSource (Maybe Text)

The label of the caption shown in the player when choosing a language. We recommend that you put the caption language name here, in the language of the captions.