Network.AWS.ElasticTranscoder.Types.DetectedProperties

Description

Documentation

data DetectedProperties

The detected properties of the input file. Elastic Transcoder identifies these values from the input file.

See: newDetectedProperties smart constructor.

Constructors

DetectedProperties'

Fields

height :: Maybe Int

The detected height of the input file, in pixels.

frameRate :: Maybe Text

The detected frame rate of the input file, in frames per second.

fileSize :: Maybe Integer

The detected file size of the input file, in bytes.

width :: Maybe Int

The detected width of the input file, in pixels.

durationMillis :: Maybe Integer

The detected duration of the input file, in milliseconds.

newDetectedProperties :: DetectedProperties

Create a value of DetectedProperties with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:height:DetectedProperties', detectedProperties_height - The detected height of the input file, in pixels.

$sel:frameRate:DetectedProperties', detectedProperties_frameRate - The detected frame rate of the input file, in frames per second.

$sel:fileSize:DetectedProperties', detectedProperties_fileSize - The detected file size of the input file, in bytes.

$sel:width:DetectedProperties', detectedProperties_width - The detected width of the input file, in pixels.

$sel:durationMillis:DetectedProperties', detectedProperties_durationMillis - The detected duration of the input file, in milliseconds.

detectedProperties_height :: Lens' DetectedProperties (Maybe Int)

The detected height of the input file, in pixels.

detectedProperties_frameRate :: Lens' DetectedProperties (Maybe Text)

The detected frame rate of the input file, in frames per second.

detectedProperties_fileSize :: Lens' DetectedProperties (Maybe Integer)

The detected file size of the input file, in bytes.

detectedProperties_width :: Lens' DetectedProperties (Maybe Int)

The detected width of the input file, in pixels.

detectedProperties_durationMillis :: Lens' DetectedProperties (Maybe Integer)

The detected duration of the input file, in milliseconds.