Network.AWS.ElasticTranscoder.Types.JobAlbumArt

Description

Documentation

data JobAlbumArt

The .jpg or .png file associated with an audio file.

See: newJobAlbumArt smart constructor.

Constructors

JobAlbumArt'

Fields

mergePolicy :: Maybe Text

A policy that determines how Elastic Transcoder handles the existence of multiple album artwork files.

Replace: The specified album art replaces any existing album art.

Prepend: The specified album art is placed in front of any existing album art.

Append: The specified album art is placed after any existing album art.

Fallback: If the original input file contains artwork, Elastic Transcoder uses that artwork for the output. If the original input does not contain artwork, Elastic Transcoder uses the specified album art file.

artwork :: Maybe [Artwork]

The file to be used as album art. There can be multiple artworks associated with an audio file, to a maximum of 20. Valid formats are .jpg and .png

newJobAlbumArt :: JobAlbumArt

Create a value of JobAlbumArt with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:mergePolicy:JobAlbumArt', jobAlbumArt_mergePolicy - A policy that determines how Elastic Transcoder handles the existence of multiple album artwork files.

Replace: The specified album art replaces any existing album art.

Prepend: The specified album art is placed in front of any existing album art.

Append: The specified album art is placed after any existing album art.

Fallback: If the original input file contains artwork, Elastic Transcoder uses that artwork for the output. If the original input does not contain artwork, Elastic Transcoder uses the specified album art file.

$sel:artwork:JobAlbumArt', jobAlbumArt_artwork - The file to be used as album art. There can be multiple artworks associated with an audio file, to a maximum of 20. Valid formats are .jpg and .png

jobAlbumArt_mergePolicy :: Lens' JobAlbumArt (Maybe Text)

A policy that determines how Elastic Transcoder handles the existence of multiple album artwork files.

Replace: The specified album art replaces any existing album art.

Prepend: The specified album art is placed in front of any existing album art.

Append: The specified album art is placed after any existing album art.

Fallback: If the original input file contains artwork, Elastic Transcoder uses that artwork for the output. If the original input does not contain artwork, Elastic Transcoder uses the specified album art file.

jobAlbumArt_artwork :: Lens' JobAlbumArt (Maybe [Artwork])

The file to be used as album art. There can be multiple artworks associated with an audio file, to a maximum of 20. Valid formats are .jpg and .png