Network.AWS.ElasticTranscoder.Types.Pipeline

Description

Documentation

data Pipeline

The pipeline (queue) that is used to manage jobs.

See: newPipeline smart constructor.

Constructors

Pipeline'

Fields

status :: Maybe Text

The current status of the pipeline:

Active: The pipeline is processing jobs.

Paused: The pipeline is not currently processing jobs.

arn :: Maybe Text

The Amazon Resource Name (ARN) for the pipeline.

inputBucket :: Maybe Text

The Amazon S3 bucket from which Elastic Transcoder gets media files for transcoding and the graphics files, if any, that you want to use for watermarks.

contentConfig :: Maybe PipelineOutputConfig

Information about the Amazon S3 bucket in which you want Elastic Transcoder to save transcoded files and playlists. Either you specify both ContentConfig and ThumbnailConfig, or you specify OutputBucket.

Bucket: The Amazon S3 bucket in which you want Elastic Transcoder to save transcoded files and playlists.

Permissions: A list of the users and/or predefined Amazon S3 groups you want to have access to transcoded files and playlists, and the type of access that you want them to have.

GranteeType: The type of value that appears in the Grantee object:

Canonical: Either the canonical user ID for an AWS account or an origin access identity for an Amazon CloudFront distribution.

Email: The registered email address of an AWS account.

Group: One of the following predefined Amazon S3 groups: AllUsers, AuthenticatedUsers, or LogDelivery.

Grantee: The AWS user or group that you want to have access to transcoded files and playlists.

Access: The permission that you want to give to the AWS user that is listed in Grantee. Valid values include:

READ: The grantee can read the objects and metadata for objects that Elastic Transcoder adds to the Amazon S3 bucket.

READ_ACP: The grantee can read the object ACL for objects that Elastic Transcoder adds to the Amazon S3 bucket.

WRITE_ACP: The grantee can write the ACL for the objects that Elastic Transcoder adds to the Amazon S3 bucket.

FULL_CONTROL: The grantee has READ, READ_ACP, and WRITE_ACP permissions for the objects that Elastic Transcoder adds to the Amazon S3 bucket.

StorageClass: The Amazon S3 storage class, Standard or ReducedRedundancy, that you want Elastic Transcoder to assign to the video files and playlists that it stores in your Amazon S3 bucket.

outputBucket :: Maybe Text

The Amazon S3 bucket in which you want Elastic Transcoder to save transcoded files, thumbnails, and playlists. Either you specify this value, or you specify both ContentConfig and ThumbnailConfig.

role' :: Maybe Text

The IAM Amazon Resource Name (ARN) for the role that Elastic Transcoder uses to transcode jobs for this pipeline.

name :: Maybe Text

The name of the pipeline. We recommend that the name be unique within the AWS account, but uniqueness is not enforced.

Constraints: Maximum 40 characters

awsKmsKeyArn :: Maybe Text

The AWS Key Management Service (AWS KMS) key that you want to use with this pipeline.

If you use either s3 or s3-aws-kms as your Encryption:Mode, you don't need to provide a key with your job because a default key, known as an AWS-KMS key, is created for you automatically. You need to provide an AWS-KMS key only if you want to use a non-default AWS-KMS key, or if you are using an Encryption:Mode of aes-cbc-pkcs7, aes-ctr, or aes-gcm.

id :: Maybe Text

The identifier for the pipeline. You use this value to identify the pipeline in which you want to perform a variety of operations, such as creating a job or a preset.

notifications :: Maybe Notifications

The Amazon Simple Notification Service (Amazon SNS) topic that you want to notify to report job status.

To receive notifications, you must also subscribe to the new topic in the Amazon SNS console.

Progressing (optional): The Amazon Simple Notification Service (Amazon SNS) topic that you want to notify when Elastic Transcoder has started to process the job.

Complete (optional): The Amazon SNS topic that you want to notify when Elastic Transcoder has finished processing the job.

Warning (optional): The Amazon SNS topic that you want to notify when Elastic Transcoder encounters a warning condition.

Error (optional): The Amazon SNS topic that you want to notify when Elastic Transcoder encounters an error condition.

thumbnailConfig :: Maybe PipelineOutputConfig

Information about the Amazon S3 bucket in which you want Elastic Transcoder to save thumbnail files. Either you specify both ContentConfig and ThumbnailConfig, or you specify OutputBucket.

Bucket: The Amazon S3 bucket in which you want Elastic Transcoder to save thumbnail files.

Permissions: A list of the users and/or predefined Amazon S3 groups you want to have access to thumbnail files, and the type of access that you want them to have.

GranteeType: The type of value that appears in the Grantee object:

Canonical: Either the canonical user ID for an AWS account or an origin access identity for an Amazon CloudFront distribution.

A canonical user ID is not the same as an AWS account number.

Email: The registered email address of an AWS account.

Group: One of the following predefined Amazon S3 groups: AllUsers, AuthenticatedUsers, or LogDelivery.

Grantee: The AWS user or group that you want to have access to thumbnail files.

Access: The permission that you want to give to the AWS user that is listed in Grantee. Valid values include:

READ: The grantee can read the thumbnails and metadata for thumbnails that Elastic Transcoder adds to the Amazon S3 bucket.

READ_ACP: The grantee can read the object ACL for thumbnails that Elastic Transcoder adds to the Amazon S3 bucket.

WRITE_ACP: The grantee can write the ACL for the thumbnails that Elastic Transcoder adds to the Amazon S3 bucket.

FULL_CONTROL: The grantee has READ, READ_ACP, and WRITE_ACP permissions for the thumbnails that Elastic Transcoder adds to the Amazon S3 bucket.

StorageClass: The Amazon S3 storage class, Standard or ReducedRedundancy, that you want Elastic Transcoder to assign to the thumbnails that it stores in your Amazon S3 bucket.

newPipeline :: Pipeline

Create a value of Pipeline with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Pipeline', pipeline_status - The current status of the pipeline:

Active: The pipeline is processing jobs.

Paused: The pipeline is not currently processing jobs.

$sel:arn:Pipeline', pipeline_arn - The Amazon Resource Name (ARN) for the pipeline.

$sel:inputBucket:Pipeline', pipeline_inputBucket - The Amazon S3 bucket from which Elastic Transcoder gets media files for transcoding and the graphics files, if any, that you want to use for watermarks.

$sel:contentConfig:Pipeline', pipeline_contentConfig - Information about the Amazon S3 bucket in which you want Elastic Transcoder to save transcoded files and playlists. Either you specify both ContentConfig and ThumbnailConfig, or you specify OutputBucket.

Bucket: The Amazon S3 bucket in which you want Elastic Transcoder to save transcoded files and playlists.

Permissions: A list of the users and/or predefined Amazon S3 groups you want to have access to transcoded files and playlists, and the type of access that you want them to have.

GranteeType: The type of value that appears in the Grantee object:

Canonical: Either the canonical user ID for an AWS account or an origin access identity for an Amazon CloudFront distribution.

Email: The registered email address of an AWS account.

Group: One of the following predefined Amazon S3 groups: AllUsers, AuthenticatedUsers, or LogDelivery.

Grantee: The AWS user or group that you want to have access to transcoded files and playlists.

Access: The permission that you want to give to the AWS user that is listed in Grantee. Valid values include:

READ: The grantee can read the objects and metadata for objects that Elastic Transcoder adds to the Amazon S3 bucket.

READ_ACP: The grantee can read the object ACL for objects that Elastic Transcoder adds to the Amazon S3 bucket.

WRITE_ACP: The grantee can write the ACL for the objects that Elastic Transcoder adds to the Amazon S3 bucket.

FULL_CONTROL: The grantee has READ, READ_ACP, and WRITE_ACP permissions for the objects that Elastic Transcoder adds to the Amazon S3 bucket.

StorageClass: The Amazon S3 storage class, Standard or ReducedRedundancy, that you want Elastic Transcoder to assign to the video files and playlists that it stores in your Amazon S3 bucket.

$sel:outputBucket:Pipeline', pipeline_outputBucket - The Amazon S3 bucket in which you want Elastic Transcoder to save transcoded files, thumbnails, and playlists. Either you specify this value, or you specify both ContentConfig and ThumbnailConfig.

$sel:role':Pipeline', pipeline_role - The IAM Amazon Resource Name (ARN) for the role that Elastic Transcoder uses to transcode jobs for this pipeline.

$sel:name:Pipeline', pipeline_name - The name of the pipeline. We recommend that the name be unique within the AWS account, but uniqueness is not enforced.

Constraints: Maximum 40 characters

$sel:awsKmsKeyArn:Pipeline', pipeline_awsKmsKeyArn - The AWS Key Management Service (AWS KMS) key that you want to use with this pipeline.

If you use either s3 or s3-aws-kms as your Encryption:Mode, you don't need to provide a key with your job because a default key, known as an AWS-KMS key, is created for you automatically. You need to provide an AWS-KMS key only if you want to use a non-default AWS-KMS key, or if you are using an Encryption:Mode of aes-cbc-pkcs7, aes-ctr, or aes-gcm.

$sel:id:Pipeline', pipeline_id - The identifier for the pipeline. You use this value to identify the pipeline in which you want to perform a variety of operations, such as creating a job or a preset.

$sel:notifications:Pipeline', pipeline_notifications - The Amazon Simple Notification Service (Amazon SNS) topic that you want to notify to report job status.

To receive notifications, you must also subscribe to the new topic in the Amazon SNS console.

Progressing (optional): The Amazon Simple Notification Service (Amazon SNS) topic that you want to notify when Elastic Transcoder has started to process the job.

Complete (optional): The Amazon SNS topic that you want to notify when Elastic Transcoder has finished processing the job.

Warning (optional): The Amazon SNS topic that you want to notify when Elastic Transcoder encounters a warning condition.

Error (optional): The Amazon SNS topic that you want to notify when Elastic Transcoder encounters an error condition.

$sel:thumbnailConfig:Pipeline', pipeline_thumbnailConfig - Information about the Amazon S3 bucket in which you want Elastic Transcoder to save thumbnail files. Either you specify both ContentConfig and ThumbnailConfig, or you specify OutputBucket.

Bucket: The Amazon S3 bucket in which you want Elastic Transcoder to save thumbnail files.

Permissions: A list of the users and/or predefined Amazon S3 groups you want to have access to thumbnail files, and the type of access that you want them to have.

GranteeType: The type of value that appears in the Grantee object:

Canonical: Either the canonical user ID for an AWS account or an origin access identity for an Amazon CloudFront distribution.

A canonical user ID is not the same as an AWS account number.

Email: The registered email address of an AWS account.

Group: One of the following predefined Amazon S3 groups: AllUsers, AuthenticatedUsers, or LogDelivery.

Grantee: The AWS user or group that you want to have access to thumbnail files.

Access: The permission that you want to give to the AWS user that is listed in Grantee. Valid values include:

READ: The grantee can read the thumbnails and metadata for thumbnails that Elastic Transcoder adds to the Amazon S3 bucket.

READ_ACP: The grantee can read the object ACL for thumbnails that Elastic Transcoder adds to the Amazon S3 bucket.

WRITE_ACP: The grantee can write the ACL for the thumbnails that Elastic Transcoder adds to the Amazon S3 bucket.

FULL_CONTROL: The grantee has READ, READ_ACP, and WRITE_ACP permissions for the thumbnails that Elastic Transcoder adds to the Amazon S3 bucket.

StorageClass: The Amazon S3 storage class, Standard or ReducedRedundancy, that you want Elastic Transcoder to assign to the thumbnails that it stores in your Amazon S3 bucket.

pipeline_status :: Lens' Pipeline (Maybe Text)

The current status of the pipeline:

Active: The pipeline is processing jobs.

Paused: The pipeline is not currently processing jobs.

pipeline_arn :: Lens' Pipeline (Maybe Text)

The Amazon Resource Name (ARN) for the pipeline.

pipeline_inputBucket :: Lens' Pipeline (Maybe Text)

The Amazon S3 bucket from which Elastic Transcoder gets media files for transcoding and the graphics files, if any, that you want to use for watermarks.

pipeline_contentConfig :: Lens' Pipeline (Maybe PipelineOutputConfig)

Information about the Amazon S3 bucket in which you want Elastic Transcoder to save transcoded files and playlists. Either you specify both ContentConfig and ThumbnailConfig, or you specify OutputBucket.

Bucket: The Amazon S3 bucket in which you want Elastic Transcoder to save transcoded files and playlists.

Permissions: A list of the users and/or predefined Amazon S3 groups you want to have access to transcoded files and playlists, and the type of access that you want them to have.

GranteeType: The type of value that appears in the Grantee object:

Canonical: Either the canonical user ID for an AWS account or an origin access identity for an Amazon CloudFront distribution.

Email: The registered email address of an AWS account.

Group: One of the following predefined Amazon S3 groups: AllUsers, AuthenticatedUsers, or LogDelivery.

Grantee: The AWS user or group that you want to have access to transcoded files and playlists.

Access: The permission that you want to give to the AWS user that is listed in Grantee. Valid values include:

READ: The grantee can read the objects and metadata for objects that Elastic Transcoder adds to the Amazon S3 bucket.

READ_ACP: The grantee can read the object ACL for objects that Elastic Transcoder adds to the Amazon S3 bucket.

WRITE_ACP: The grantee can write the ACL for the objects that Elastic Transcoder adds to the Amazon S3 bucket.

FULL_CONTROL: The grantee has READ, READ_ACP, and WRITE_ACP permissions for the objects that Elastic Transcoder adds to the Amazon S3 bucket.

StorageClass: The Amazon S3 storage class, Standard or ReducedRedundancy, that you want Elastic Transcoder to assign to the video files and playlists that it stores in your Amazon S3 bucket.

pipeline_outputBucket :: Lens' Pipeline (Maybe Text)

The Amazon S3 bucket in which you want Elastic Transcoder to save transcoded files, thumbnails, and playlists. Either you specify this value, or you specify both ContentConfig and ThumbnailConfig.

pipeline_role :: Lens' Pipeline (Maybe Text)

The IAM Amazon Resource Name (ARN) for the role that Elastic Transcoder uses to transcode jobs for this pipeline.

pipeline_name :: Lens' Pipeline (Maybe Text)

The name of the pipeline. We recommend that the name be unique within the AWS account, but uniqueness is not enforced.

Constraints: Maximum 40 characters

pipeline_awsKmsKeyArn :: Lens' Pipeline (Maybe Text)

The AWS Key Management Service (AWS KMS) key that you want to use with this pipeline.

If you use either s3 or s3-aws-kms as your Encryption:Mode, you don't need to provide a key with your job because a default key, known as an AWS-KMS key, is created for you automatically. You need to provide an AWS-KMS key only if you want to use a non-default AWS-KMS key, or if you are using an Encryption:Mode of aes-cbc-pkcs7, aes-ctr, or aes-gcm.

pipeline_id :: Lens' Pipeline (Maybe Text)

The identifier for the pipeline. You use this value to identify the pipeline in which you want to perform a variety of operations, such as creating a job or a preset.

pipeline_notifications :: Lens' Pipeline (Maybe Notifications)

The Amazon Simple Notification Service (Amazon SNS) topic that you want to notify to report job status.

To receive notifications, you must also subscribe to the new topic in the Amazon SNS console.

Progressing (optional): The Amazon Simple Notification Service (Amazon SNS) topic that you want to notify when Elastic Transcoder has started to process the job.

Complete (optional): The Amazon SNS topic that you want to notify when Elastic Transcoder has finished processing the job.

Warning (optional): The Amazon SNS topic that you want to notify when Elastic Transcoder encounters a warning condition.

Error (optional): The Amazon SNS topic that you want to notify when Elastic Transcoder encounters an error condition.

pipeline_thumbnailConfig :: Lens' Pipeline (Maybe PipelineOutputConfig)

Information about the Amazon S3 bucket in which you want Elastic Transcoder to save thumbnail files. Either you specify both ContentConfig and ThumbnailConfig, or you specify OutputBucket.

Bucket: The Amazon S3 bucket in which you want Elastic Transcoder to save thumbnail files.

Permissions: A list of the users and/or predefined Amazon S3 groups you want to have access to thumbnail files, and the type of access that you want them to have.

GranteeType: The type of value that appears in the Grantee object:

Canonical: Either the canonical user ID for an AWS account or an origin access identity for an Amazon CloudFront distribution.

A canonical user ID is not the same as an AWS account number.

Email: The registered email address of an AWS account.

Group: One of the following predefined Amazon S3 groups: AllUsers, AuthenticatedUsers, or LogDelivery.

Grantee: The AWS user or group that you want to have access to thumbnail files.

Access: The permission that you want to give to the AWS user that is listed in Grantee. Valid values include:

READ: The grantee can read the thumbnails and metadata for thumbnails that Elastic Transcoder adds to the Amazon S3 bucket.

READ_ACP: The grantee can read the object ACL for thumbnails that Elastic Transcoder adds to the Amazon S3 bucket.

WRITE_ACP: The grantee can write the ACL for the thumbnails that Elastic Transcoder adds to the Amazon S3 bucket.

FULL_CONTROL: The grantee has READ, READ_ACP, and WRITE_ACP permissions for the thumbnails that Elastic Transcoder adds to the Amazon S3 bucket.

StorageClass: The Amazon S3 storage class, Standard or ReducedRedundancy, that you want Elastic Transcoder to assign to the thumbnails that it stores in your Amazon S3 bucket.