Network.AWS.ElasticTranscoder.Types.Playlist

Description

Documentation

data Playlist

Use Only for Fragmented MP4 or MPEG-TS Outputs. If you specify a preset for which the value of Container is fmp4 (Fragmented MP4) or ts (MPEG-TS), Playlists contains information about the master playlists that you want Elastic Transcoder to create. We recommend that you create only one master playlist per output format. The maximum number of master playlists in a job is 30.

See: newPlaylist smart constructor.

Constructors

Playlist'

Fields

status :: Maybe Text

The status of the job with which the playlist is associated.

playReadyDrm :: Maybe PlayReadyDrm

The DRM settings, if any, that you want Elastic Transcoder to apply to the output files associated with this playlist.

format :: Maybe Text

The format of the output playlist. Valid formats include HLSv3, HLSv4, and Smooth.

outputKeys :: Maybe [Text]

For each output in this job that you want to include in a master playlist, the value of the Outputs:Key object.

If your output is not HLS or does not have a segment duration set, the name of the output file is a concatenation of OutputKeyPrefix and Outputs:Key:

OutputKeyPrefixOutputs:Key

If your output is HLSv3 and has a segment duration set, or is not included in a playlist, Elastic Transcoder creates an output playlist file with a file extension of .m3u8, and a series of .ts files that include a five-digit sequential counter beginning with 00000:

OutputKeyPrefixOutputs:Key.m3u8

OutputKeyPrefixOutputs:Key00000.ts

If your output is HLSv4, has a segment duration set, and is included in an HLSv4 playlist, Elastic Transcoder creates an output playlist file with a file extension of _v4.m3u8. If the output is video, Elastic Transcoder also creates an output file with an extension of _iframe.m3u8:

OutputKeyPrefixOutputs:Key_v4.m3u8

OutputKeyPrefixOutputs:Key_iframe.m3u8

OutputKeyPrefixOutputs:Key.ts

Elastic Transcoder automatically appends the relevant file extension to the file name. If you include a file extension in Output Key, the file name will have two extensions.

If you include more than one output in a playlist, any segment duration settings, clip settings, or caption settings must be the same for all outputs in the playlist. For Smooth playlists, the Audio:Profile, Video:Profile, and Video:FrameRate to Video:KeyframesMaxDist ratio must be the same for all outputs.

name :: Maybe Text

The name that you want Elastic Transcoder to assign to the master playlist, for example, nyc-vacation.m3u8. If the name includes a / character, the section of the name before the last / must be identical for all Name objects. If you create more than one master playlist, the values of all Name objects must be unique.

Elastic Transcoder automatically appends the relevant file extension to the file name (.m3u8 for HLSv3 and HLSv4 playlists, and .ism and .ismc for Smooth playlists). If you include a file extension in Name, the file name will have two extensions.

statusDetail :: Maybe Text

Information that further explains the status.

hlsContentProtection :: Maybe HlsContentProtection

The HLS content protection settings, if any, that you want Elastic Transcoder to apply to the output files associated with this playlist.

newPlaylist :: Playlist

Create a value of Playlist with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Playlist', playlist_status - The status of the job with which the playlist is associated.

$sel:playReadyDrm:Playlist', playlist_playReadyDrm - The DRM settings, if any, that you want Elastic Transcoder to apply to the output files associated with this playlist.

$sel:format:Playlist', playlist_format - The format of the output playlist. Valid formats include HLSv3, HLSv4, and Smooth.

$sel:outputKeys:Playlist', playlist_outputKeys - For each output in this job that you want to include in a master playlist, the value of the Outputs:Key object.

If your output is not HLS or does not have a segment duration set, the name of the output file is a concatenation of OutputKeyPrefix and Outputs:Key:

OutputKeyPrefixOutputs:Key

If your output is HLSv3 and has a segment duration set, or is not included in a playlist, Elastic Transcoder creates an output playlist file with a file extension of .m3u8, and a series of .ts files that include a five-digit sequential counter beginning with 00000:

OutputKeyPrefixOutputs:Key.m3u8

OutputKeyPrefixOutputs:Key00000.ts

If your output is HLSv4, has a segment duration set, and is included in an HLSv4 playlist, Elastic Transcoder creates an output playlist file with a file extension of _v4.m3u8. If the output is video, Elastic Transcoder also creates an output file with an extension of _iframe.m3u8:

OutputKeyPrefixOutputs:Key_v4.m3u8

OutputKeyPrefixOutputs:Key_iframe.m3u8

OutputKeyPrefixOutputs:Key.ts

Elastic Transcoder automatically appends the relevant file extension to the file name. If you include a file extension in Output Key, the file name will have two extensions.

If you include more than one output in a playlist, any segment duration settings, clip settings, or caption settings must be the same for all outputs in the playlist. For Smooth playlists, the Audio:Profile, Video:Profile, and Video:FrameRate to Video:KeyframesMaxDist ratio must be the same for all outputs.

$sel:name:Playlist', playlist_name - The name that you want Elastic Transcoder to assign to the master playlist, for example, nyc-vacation.m3u8. If the name includes a / character, the section of the name before the last / must be identical for all Name objects. If you create more than one master playlist, the values of all Name objects must be unique.

Elastic Transcoder automatically appends the relevant file extension to the file name (.m3u8 for HLSv3 and HLSv4 playlists, and .ism and .ismc for Smooth playlists). If you include a file extension in Name, the file name will have two extensions.

$sel:statusDetail:Playlist', playlist_statusDetail - Information that further explains the status.

$sel:hlsContentProtection:Playlist', playlist_hlsContentProtection - The HLS content protection settings, if any, that you want Elastic Transcoder to apply to the output files associated with this playlist.

playlist_status :: Lens' Playlist (Maybe Text)

The status of the job with which the playlist is associated.

playlist_playReadyDrm :: Lens' Playlist (Maybe PlayReadyDrm)

The DRM settings, if any, that you want Elastic Transcoder to apply to the output files associated with this playlist.

playlist_format :: Lens' Playlist (Maybe Text)

The format of the output playlist. Valid formats include HLSv3, HLSv4, and Smooth.

playlist_outputKeys :: Lens' Playlist (Maybe [Text])

For each output in this job that you want to include in a master playlist, the value of the Outputs:Key object.

If your output is not HLS or does not have a segment duration set, the name of the output file is a concatenation of OutputKeyPrefix and Outputs:Key:

OutputKeyPrefixOutputs:Key

If your output is HLSv3 and has a segment duration set, or is not included in a playlist, Elastic Transcoder creates an output playlist file with a file extension of .m3u8, and a series of .ts files that include a five-digit sequential counter beginning with 00000:

OutputKeyPrefixOutputs:Key.m3u8

OutputKeyPrefixOutputs:Key00000.ts

If your output is HLSv4, has a segment duration set, and is included in an HLSv4 playlist, Elastic Transcoder creates an output playlist file with a file extension of _v4.m3u8. If the output is video, Elastic Transcoder also creates an output file with an extension of _iframe.m3u8:

OutputKeyPrefixOutputs:Key_v4.m3u8

OutputKeyPrefixOutputs:Key_iframe.m3u8

OutputKeyPrefixOutputs:Key.ts

Elastic Transcoder automatically appends the relevant file extension to the file name. If you include a file extension in Output Key, the file name will have two extensions.

If you include more than one output in a playlist, any segment duration settings, clip settings, or caption settings must be the same for all outputs in the playlist. For Smooth playlists, the Audio:Profile, Video:Profile, and Video:FrameRate to Video:KeyframesMaxDist ratio must be the same for all outputs.

playlist_name :: Lens' Playlist (Maybe Text)

The name that you want Elastic Transcoder to assign to the master playlist, for example, nyc-vacation.m3u8. If the name includes a / character, the section of the name before the last / must be identical for all Name objects. If you create more than one master playlist, the values of all Name objects must be unique.

Elastic Transcoder automatically appends the relevant file extension to the file name (.m3u8 for HLSv3 and HLSv4 playlists, and .ism and .ismc for Smooth playlists). If you include a file extension in Name, the file name will have two extensions.

playlist_statusDetail :: Lens' Playlist (Maybe Text)

Information that further explains the status.

playlist_hlsContentProtection :: Lens' Playlist (Maybe HlsContentProtection)

The HLS content protection settings, if any, that you want Elastic Transcoder to apply to the output files associated with this playlist.