Network.AWS.ElasticTranscoder.Types.Preset

Description

Documentation

data Preset

Presets are templates that contain most of the settings for transcoding media files from one format to another. Elastic Transcoder includes some default presets for common formats, for example, several iPod and iPhone versions. You can also create your own presets for formats that aren't included among the default presets. You specify which preset you want to use when you create a job.

See: newPreset smart constructor.

Constructors

Preset'

Fields

arn :: Maybe Text

The Amazon Resource Name (ARN) for the preset.

video :: Maybe VideoParameters

A section of the response body that provides information about the video preset values.

thumbnails :: Maybe Thumbnails

A section of the response body that provides information about the thumbnail preset values, if any.

name :: Maybe Text

The name of the preset.

container :: Maybe Text

The container type for the output file. Valid values include flac, flv, fmp4, gif, mp3, mp4, mpg, mxf, oga, ogg, ts, and webm.

id :: Maybe Text

Identifier for the new preset. You use this value to get settings for the preset or to delete it.

type' :: Maybe Text

Whether the preset is a default preset provided by Elastic Transcoder (System) or a preset that you have defined (Custom).

description :: Maybe Text

A description of the preset.

audio :: Maybe AudioParameters

A section of the response body that provides information about the audio preset values.

newPreset :: Preset

Create a value of Preset with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:Preset', preset_arn - The Amazon Resource Name (ARN) for the preset.

$sel:video:Preset', preset_video - A section of the response body that provides information about the video preset values.

$sel:thumbnails:Preset', preset_thumbnails - A section of the response body that provides information about the thumbnail preset values, if any.

$sel:name:Preset', preset_name - The name of the preset.

$sel:container:Preset', preset_container - The container type for the output file. Valid values include flac, flv, fmp4, gif, mp3, mp4, mpg, mxf, oga, ogg, ts, and webm.

$sel:id:Preset', preset_id - Identifier for the new preset. You use this value to get settings for the preset or to delete it.

$sel:type':Preset', preset_type - Whether the preset is a default preset provided by Elastic Transcoder (System) or a preset that you have defined (Custom).

$sel:description:Preset', preset_description - A description of the preset.

$sel:audio:Preset', preset_audio - A section of the response body that provides information about the audio preset values.

preset_arn :: Lens' Preset (Maybe Text)

The Amazon Resource Name (ARN) for the preset.

preset_video :: Lens' Preset (Maybe VideoParameters)

A section of the response body that provides information about the video preset values.

preset_thumbnails :: Lens' Preset (Maybe Thumbnails)

A section of the response body that provides information about the thumbnail preset values, if any.

preset_name :: Lens' Preset (Maybe Text)

The name of the preset.

preset_container :: Lens' Preset (Maybe Text)

The container type for the output file. Valid values include flac, flv, fmp4, gif, mp3, mp4, mpg, mxf, oga, ogg, ts, and webm.

preset_id :: Lens' Preset (Maybe Text)

Identifier for the new preset. You use this value to get settings for the preset or to delete it.

preset_type :: Lens' Preset (Maybe Text)

Whether the preset is a default preset provided by Elastic Transcoder (System) or a preset that you have defined (Custom).

preset_description :: Lens' Preset (Maybe Text)

A description of the preset.

preset_audio :: Lens' Preset (Maybe AudioParameters)

A section of the response body that provides information about the audio preset values.