Network.AWS.ElasticTranscoder.Types.Timing

Description

Documentation

data Timing

Details about the timing of a job.

See: newTiming smart constructor.

Constructors

Timing'

Fields

submitTimeMillis :: Maybe Integer

The time the job was submitted to Elastic Transcoder, in epoch milliseconds.

finishTimeMillis :: Maybe Integer

The time the job finished transcoding, in epoch milliseconds.

startTimeMillis :: Maybe Integer

The time the job began transcoding, in epoch milliseconds.

newTiming :: Timing

Create a value of Timing with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:submitTimeMillis:Timing', timing_submitTimeMillis - The time the job was submitted to Elastic Transcoder, in epoch milliseconds.

$sel:finishTimeMillis:Timing', timing_finishTimeMillis - The time the job finished transcoding, in epoch milliseconds.

$sel:startTimeMillis:Timing', timing_startTimeMillis - The time the job began transcoding, in epoch milliseconds.

timing_submitTimeMillis :: Lens' Timing (Maybe Integer)

The time the job was submitted to Elastic Transcoder, in epoch milliseconds.

timing_finishTimeMillis :: Lens' Timing (Maybe Integer)

The time the job finished transcoding, in epoch milliseconds.

timing_startTimeMillis :: Lens' Timing (Maybe Integer)

The time the job began transcoding, in epoch milliseconds.