Network.AWS.ElasticTranscoder.V2012_09_25.DeletePreset

Description

The DeletePreset operation removes a preset that you've added in an AWS region. You can't delete the default presets that are included with Elastic Transcoder. DELETE 2012-09-25pipelines5555555555555-abcde5 HTTP1.1 Content-Type: charset=UTF-8 Accept: */* Host: elastictranscoder.[Elastic Transcoder-endpoint].amazonaws.com:443 x-amz-date: 20130114T174952Z Authorization: AWS4-HMAC-SHA256 Credential=[access-key-id][request-date][Elastic Transcoder-endpoint]etsaws4_request, SignedHeaders=host;x-amz-date;x-amz-target, Signature=[calculated-signature] Status: 202 Accepted x-amzn-RequestId: c321ec43-378e-11e2-8e4c-4d5b971203e9 Content-Type: application/json Content-Length: [number-of-characters-in-response] Date: Mon, 14 Jan 2013 06:01:47 GMT { Success:"true" }.

Documentation

data DeletePreset

Constructors

DeletePreset

Fields

_dprId :: Text

The identifier of the preset for which you want to get detailed information.

dprId :: Iso' DeletePreset Text

data DeletePresetResponse

Constructors

DeletePresetResponse