Network.AWS.ElasticTranscoder.V2012_09_25.ReadPreset

Description

The ReadPreset operation gets detailed information about a preset. GET 2012-09-25presets5555555555555-abcde5 HTTP1.1 Content-Type: applicationjson; charset=UTF-8 Accept: ** Host: elastictranscoder.[Elastic Transcoder-endpoint].amazonaws.com:443 x-amz-date: 20130114T174952Z Authorization: AWS4-HMAC-SHA256 Credential=[access-key-id][request-date][Elastic Transcoder-endpoint]etsaws4_request, SignedHeaders=host;x-amz-date;x-amz-target, Signature=[calculated-signature] Content-Length: [number-of-characters-in-JSON-string] Status: 200 OK Content-Type: charset=UTF-8 Accept: */* Host: elastictranscoder.[Elastic Transcoder-endpoint].amazonaws.com:443 x-amz-date: 20130114T174952Z Authorization: AWS4-HMAC-SHA256 Credential=[access-key-id][request-date][Elastic Transcoder-endpoint]etsaws4_request, SignedHeaders=host;x-amz-date;x-amz-target, Signature=[calculated-signature] { Preset:{ Audio:{ Codec:AAC, SampleRate:"44100", BitRate:96, Channels:2 }, Container:"mp4", Description:"Use for published videos", Id:"5555555555555-abcde5", Name:DefaultPreset, Thumbnails:{ Format:"png", Interval:"120", MaxWidth: "auto", MaxHeight: "auto", SizingPolicy: Fill, PaddingPolicy: Pad }, Type:Custom, Video:{ MaxWidth: "auto", MaxHeight: "auto", SizingPolicy: Fill, PaddingPolicy: Pad, DisplayAspectRatio: "auto", BitRate:"1600", Codec:"H.264", CodecOptions:{ Level:"2.2", MaxReferenceFrames:"3", Profile:"main", MaxBitRate:"", BufferSize:"" }, FixedGOP:"false", FrameRate:"30", KeyframesMaxDist:"240" } } }.

Documentation

data ReadPreset

Constructors

ReadPreset

Fields

_rprId :: Text

The identifier of the preset for which you want to get detailed information.

rprId :: Iso' ReadPreset Text

data ReadPresetResponse

Constructors

ReadPresetResponse

Fields

_rpsPreset :: Maybe Preset

A section of the response body that provides information about the preset.

rpsPreset :: Iso' ReadPresetResponse (Maybe Preset)