Network.AWS.Env

Description

Environment and AWS specific configuration for the AWS and AWST monads.

Creating the Environment

newEnv

Arguments

:: MonadIO m

=> Credentials
Credential discovery mechanism.

-> m Env

Creates a new environment with a new Manager without debug logging and uses getAuth to expand/discover the supplied Credentials. Lenses from AWSEnv can be used to further configure the resulting Env.

Since: 1.5.0 - The region is now retrieved from the AWS_REGION environment variable (identical to official SDKs), or defaults to us-east-1. You can override the Env region by using $sel:envRegion:Env, or the current operation's region by using within.

Since: 1.3.6 - The default logic for retrying HttpExceptions now uses retryConnectionFailure to retry specific connection failure conditions up to 3 times. Previously only service specific errors were automatically retried. This can be reverted to the old behaviour by resetting the Env using $sel:envRetryCheck:Env lens to (\_ _ -> False).

Throws AuthError when environment variables or IAM profiles cannot be read.

See: newEnvWith.

newEnvWith

Arguments

:: MonadIO m

=> Credentials
Credential discovery mechanism.

-> Manager
Preload the EC2 instance check.

-> m Env

See: newEnv

The Maybe Bool parameter is used by the EC2 instance check. By passing a value of Nothing, the check will be performed. Just True would cause the check to be skipped and the host treated as an EC2 instance.

Throws AuthError when environment variables or IAM profiles cannot be read.

data Env

The environment containing the parameters required to make AWS requests.

Constructors

Env

Fields

envRegion :: Region

envLogger :: Logger

envRetryCheck :: Int -> HttpException -> Bool

envOverride :: Dual (Endo Service)

envManager :: Manager

envAuth :: Auth

Overriding Default Configuration

override :: (Service -> Service) -> Env -> Env

Provide a function which will be added to the existing stack of overrides applied to all service configurations.

configure :: Service -> Env -> Env

Configure a specific service. All requests belonging to the supplied service will use this configuration instead of the default.

It's suggested you modify the default service configuration, such as Network.AWS.DynamoDB.dynamoDB.

Scoped Actions

within :: Region -> Env -> Env

Scope an action within the specific Region.

once :: Env -> Env

Scope an action such that any retry logic for the Service is ignored and any requests will at most be sent once.

timeout :: Seconds -> Env -> Env

Scope an action such that any HTTP response will use this timeout value.

Default timeouts are chosen by considering:

This timeout, if set.

The related Service timeout for the sent request if set. (Usually 70s)

The $sel:envManager:Env timeout if set.

The default ClientRequest timeout. (Approximately 30s)

Retry HTTP Exceptions

retryConnectionFailure :: Int -> Int -> HttpException -> Bool

Retry the subset of transport specific errors encompassing connection failure up to the specific number of times.