Network.AWS.FIS.ListExperimentTemplates

Description

Lists your experiment templates.

Creating a Request

data ListExperimentTemplates

See: newListExperimentTemplates smart constructor.

Constructors

ListExperimentTemplates'

Fields

nextToken :: Maybe Text

The token for the next page of results.

maxResults :: Maybe Natural

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

newListExperimentTemplates :: ListExperimentTemplates

Create a value of ListExperimentTemplates with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListExperimentTemplates', listExperimentTemplates_nextToken - The token for the next page of results.

$sel:maxResults:ListExperimentTemplates', listExperimentTemplates_maxResults - The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

Request Lenses

listExperimentTemplates_nextToken :: Lens' ListExperimentTemplates (Maybe Text)

The token for the next page of results.

listExperimentTemplates_maxResults :: Lens' ListExperimentTemplates (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

Destructuring the Response

data ListExperimentTemplatesResponse

See: newListExperimentTemplatesResponse smart constructor.

Constructors

ListExperimentTemplatesResponse'

Fields

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

experimentTemplates :: Maybe [ExperimentTemplateSummary]

The experiment templates.

httpStatus :: Int

The response's http status code.

newListExperimentTemplatesResponse

Arguments

:: Int
$sel:httpStatus:ListExperimentTemplatesResponse'

-> ListExperimentTemplatesResponse

Create a value of ListExperimentTemplatesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListExperimentTemplates', listExperimentTemplatesResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:experimentTemplates:ListExperimentTemplatesResponse', listExperimentTemplatesResponse_experimentTemplates - The experiment templates.

$sel:httpStatus:ListExperimentTemplatesResponse', listExperimentTemplatesResponse_httpStatus - The response's http status code.

Response Lenses

listExperimentTemplatesResponse_nextToken :: Lens' ListExperimentTemplatesResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

listExperimentTemplatesResponse_experimentTemplates :: Lens' ListExperimentTemplatesResponse (Maybe [ExperimentTemplateSummary])

The experiment templates.

listExperimentTemplatesResponse_httpStatus :: Lens' ListExperimentTemplatesResponse Int

The response's http status code.