Network.AWS.FMS.DeleteNotificationChannel

Description

Deletes an Firewall Manager association with the IAM role and the Amazon Simple Notification Service (SNS) topic that is used to record Firewall Manager SNS logs.

Creating a Request

data DeleteNotificationChannel

See: newDeleteNotificationChannel smart constructor.

Constructors

DeleteNotificationChannel'

newDeleteNotificationChannel :: DeleteNotificationChannel

Create a value of DeleteNotificationChannel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

Destructuring the Response

data DeleteNotificationChannelResponse

See: newDeleteNotificationChannelResponse smart constructor.

Constructors

DeleteNotificationChannelResponse'

newDeleteNotificationChannelResponse :: DeleteNotificationChannelResponse

Create a value of DeleteNotificationChannelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.