Network.AWS.FMS.GetAdminAccount

Description

Returns the Organizations account that is associated with Firewall Manager as the Firewall Manager administrator.

Creating a Request

data GetAdminAccount

See: newGetAdminAccount smart constructor.

Constructors

GetAdminAccount'

newGetAdminAccount :: GetAdminAccount

Create a value of GetAdminAccount with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

Destructuring the Response

data GetAdminAccountResponse

See: newGetAdminAccountResponse smart constructor.

Constructors

GetAdminAccountResponse'

Fields

adminAccount :: Maybe Text

The Amazon Web Services account that is set as the Firewall Manager administrator.

roleStatus :: Maybe AccountRoleStatus

The status of the Amazon Web Services account that you set as the Firewall Manager administrator.

httpStatus :: Int

The response's http status code.

newGetAdminAccountResponse

Arguments

:: Int
$sel:httpStatus:GetAdminAccountResponse'

-> GetAdminAccountResponse

Create a value of GetAdminAccountResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:adminAccount:GetAdminAccountResponse', getAdminAccountResponse_adminAccount - The Amazon Web Services account that is set as the Firewall Manager administrator.

$sel:roleStatus:GetAdminAccountResponse', getAdminAccountResponse_roleStatus - The status of the Amazon Web Services account that you set as the Firewall Manager administrator.

$sel:httpStatus:GetAdminAccountResponse', getAdminAccountResponse_httpStatus - The response's http status code.

Response Lenses

getAdminAccountResponse_adminAccount :: Lens' GetAdminAccountResponse (Maybe Text)

The Amazon Web Services account that is set as the Firewall Manager administrator.

getAdminAccountResponse_roleStatus :: Lens' GetAdminAccountResponse (Maybe AccountRoleStatus)

The status of the Amazon Web Services account that you set as the Firewall Manager administrator.

getAdminAccountResponse_httpStatus :: Lens' GetAdminAccountResponse Int

The response's http status code.