Network.AWS.FMS.ListAppsLists

Description

Returns an array of AppsListDataSummary objects.

This operation returns paginated results.

Creating a Request

data ListAppsLists

See: newListAppsLists smart constructor.

Constructors

ListAppsLists'

Fields

nextToken :: Maybe Text

If you specify a value for MaxResults in your list request, and you have more objects than the maximum, Firewall Manager returns this token in the response. For all but the first request, you provide the token returned by the prior request in the request parameters, to retrieve the next batch of objects.

defaultLists :: Maybe Bool

Specifies whether the lists to retrieve are default lists owned by Firewall Manager.

maxResults :: Natural

The maximum number of objects that you want Firewall Manager to return for this request. If more objects are available, in the response, Firewall Manager provides a NextToken value that you can use in a subsequent call to get the next batch of objects.

If you don't specify this, Firewall Manager returns all available objects.

newListAppsLists

Arguments

:: Natural
$sel:maxResults:ListAppsLists'

-> ListAppsLists

Create a value of ListAppsLists with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAppsLists', listAppsLists_nextToken - If you specify a value for MaxResults in your list request, and you have more objects than the maximum, Firewall Manager returns this token in the response. For all but the first request, you provide the token returned by the prior request in the request parameters, to retrieve the next batch of objects.

$sel:defaultLists:ListAppsLists', listAppsLists_defaultLists - Specifies whether the lists to retrieve are default lists owned by Firewall Manager.

$sel:maxResults:ListAppsLists', listAppsLists_maxResults - The maximum number of objects that you want Firewall Manager to return for this request. If more objects are available, in the response, Firewall Manager provides a NextToken value that you can use in a subsequent call to get the next batch of objects.

If you don't specify this, Firewall Manager returns all available objects.

Request Lenses

listAppsLists_nextToken :: Lens' ListAppsLists (Maybe Text)

If you specify a value for MaxResults in your list request, and you have more objects than the maximum, Firewall Manager returns this token in the response. For all but the first request, you provide the token returned by the prior request in the request parameters, to retrieve the next batch of objects.

listAppsLists_defaultLists :: Lens' ListAppsLists (Maybe Bool)

Specifies whether the lists to retrieve are default lists owned by Firewall Manager.

listAppsLists_maxResults :: Lens' ListAppsLists Natural

The maximum number of objects that you want Firewall Manager to return for this request. If more objects are available, in the response, Firewall Manager provides a NextToken value that you can use in a subsequent call to get the next batch of objects.

If you don't specify this, Firewall Manager returns all available objects.

Destructuring the Response

data ListAppsListsResponse

See: newListAppsListsResponse smart constructor.

Constructors

ListAppsListsResponse'

Fields

nextToken :: Maybe Text

If you specify a value for MaxResults in your list request, and you have more objects than the maximum, Firewall Manager returns this token in the response. You can use this token in subsequent requests to retrieve the next batch of objects.

appsLists :: Maybe [AppsListDataSummary]

An array of AppsListDataSummary objects.

httpStatus :: Int

The response's http status code.

newListAppsListsResponse

Arguments

:: Int
$sel:httpStatus:ListAppsListsResponse'

-> ListAppsListsResponse

Create a value of ListAppsListsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAppsLists', listAppsListsResponse_nextToken - If you specify a value for MaxResults in your list request, and you have more objects than the maximum, Firewall Manager returns this token in the response. You can use this token in subsequent requests to retrieve the next batch of objects.

$sel:appsLists:ListAppsListsResponse', listAppsListsResponse_appsLists - An array of AppsListDataSummary objects.

$sel:httpStatus:ListAppsListsResponse', listAppsListsResponse_httpStatus - The response's http status code.

Response Lenses

listAppsListsResponse_nextToken :: Lens' ListAppsListsResponse (Maybe Text)

If you specify a value for MaxResults in your list request, and you have more objects than the maximum, Firewall Manager returns this token in the response. You can use this token in subsequent requests to retrieve the next batch of objects.

listAppsListsResponse_appsLists :: Lens' ListAppsListsResponse (Maybe [AppsListDataSummary])

An array of AppsListDataSummary objects.

listAppsListsResponse_httpStatus :: Lens' ListAppsListsResponse Int

The response's http status code.