Network.AWS.FMS.ListMemberAccounts

Description

Returns a MemberAccounts object that lists the member accounts in the administrator's Amazon Web Services organization.

The ListMemberAccounts must be submitted by the account that is set as the Firewall Manager administrator.

This operation returns paginated results.

Creating a Request

data ListMemberAccounts

See: newListMemberAccounts smart constructor.

Constructors

ListMemberAccounts'

Fields

nextToken :: Maybe Text

If you specify a value for MaxResults and you have more account IDs than the number that you specify for MaxResults, Firewall Manager returns a NextToken value in the response that allows you to list another group of IDs. For the second and subsequent ListMemberAccountsRequest requests, specify the value of NextToken from the previous response to get information about another batch of member account IDs.

maxResults :: Maybe Natural

Specifies the number of member account IDs that you want Firewall Manager to return for this request. If you have more IDs than the number that you specify for MaxResults, the response includes a NextToken value that you can use to get another batch of member account IDs.

newListMemberAccounts :: ListMemberAccounts

Create a value of ListMemberAccounts with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListMemberAccounts', listMemberAccounts_nextToken - If you specify a value for MaxResults and you have more account IDs than the number that you specify for MaxResults, Firewall Manager returns a NextToken value in the response that allows you to list another group of IDs. For the second and subsequent ListMemberAccountsRequest requests, specify the value of NextToken from the previous response to get information about another batch of member account IDs.

$sel:maxResults:ListMemberAccounts', listMemberAccounts_maxResults - Specifies the number of member account IDs that you want Firewall Manager to return for this request. If you have more IDs than the number that you specify for MaxResults, the response includes a NextToken value that you can use to get another batch of member account IDs.

Request Lenses

listMemberAccounts_nextToken :: Lens' ListMemberAccounts (Maybe Text)

If you specify a value for MaxResults and you have more account IDs than the number that you specify for MaxResults, Firewall Manager returns a NextToken value in the response that allows you to list another group of IDs. For the second and subsequent ListMemberAccountsRequest requests, specify the value of NextToken from the previous response to get information about another batch of member account IDs.

listMemberAccounts_maxResults :: Lens' ListMemberAccounts (Maybe Natural)

Specifies the number of member account IDs that you want Firewall Manager to return for this request. If you have more IDs than the number that you specify for MaxResults, the response includes a NextToken value that you can use to get another batch of member account IDs.

Destructuring the Response

data ListMemberAccountsResponse

See: newListMemberAccountsResponse smart constructor.

Constructors

ListMemberAccountsResponse'

Fields

nextToken :: Maybe Text

If you have more member account IDs than the number that you specified for MaxResults in the request, the response includes a NextToken value. To list more IDs, submit another ListMemberAccounts request, and specify the NextToken value from the response in the NextToken value in the next request.

memberAccounts :: Maybe [Text]

An array of account IDs.

httpStatus :: Int

The response's http status code.

newListMemberAccountsResponse

Arguments

:: Int
$sel:httpStatus:ListMemberAccountsResponse'

-> ListMemberAccountsResponse

Create a value of ListMemberAccountsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListMemberAccounts', listMemberAccountsResponse_nextToken - If you have more member account IDs than the number that you specified for MaxResults in the request, the response includes a NextToken value. To list more IDs, submit another ListMemberAccounts request, and specify the NextToken value from the response in the NextToken value in the next request.

$sel:memberAccounts:ListMemberAccountsResponse', listMemberAccountsResponse_memberAccounts - An array of account IDs.

$sel:httpStatus:ListMemberAccountsResponse', listMemberAccountsResponse_httpStatus - The response's http status code.

Response Lenses

listMemberAccountsResponse_nextToken :: Lens' ListMemberAccountsResponse (Maybe Text)

If you have more member account IDs than the number that you specified for MaxResults in the request, the response includes a NextToken value. To list more IDs, submit another ListMemberAccounts request, and specify the NextToken value from the response in the NextToken value in the next request.

listMemberAccountsResponse_memberAccounts :: Lens' ListMemberAccountsResponse (Maybe [Text])

An array of account IDs.

listMemberAccountsResponse_httpStatus :: Lens' ListMemberAccountsResponse Int

The response's http status code.