Network.AWS.FMS.PutNotificationChannel

Description

Designates the IAM role and Amazon Simple Notification Service (SNS) topic that Firewall Manager uses to record SNS logs.

To perform this action outside of the console, you must configure the SNS topic to allow the Firewall Manager role AWSServiceRoleForFMS to publish SNS logs. For more information, see Firewall Manager required permissions for API actions in the Firewall Manager Developer Guide.

Creating a Request

data PutNotificationChannel

See: newPutNotificationChannel smart constructor.

Constructors

PutNotificationChannel'

Fields

snsTopicArn :: Text

The Amazon Resource Name (ARN) of the SNS topic that collects notifications from Firewall Manager.

snsRoleName :: Text

The Amazon Resource Name (ARN) of the IAM role that allows Amazon SNS to record Firewall Manager activity.

newPutNotificationChannel

Arguments

:: Text
$sel:snsTopicArn:PutNotificationChannel'

-> Text
$sel:snsRoleName:PutNotificationChannel'

-> PutNotificationChannel

Create a value of PutNotificationChannel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:snsTopicArn:PutNotificationChannel', putNotificationChannel_snsTopicArn - The Amazon Resource Name (ARN) of the SNS topic that collects notifications from Firewall Manager.

$sel:snsRoleName:PutNotificationChannel', putNotificationChannel_snsRoleName - The Amazon Resource Name (ARN) of the IAM role that allows Amazon SNS to record Firewall Manager activity.

Request Lenses

putNotificationChannel_snsTopicArn :: Lens' PutNotificationChannel Text

The Amazon Resource Name (ARN) of the SNS topic that collects notifications from Firewall Manager.

putNotificationChannel_snsRoleName :: Lens' PutNotificationChannel Text

The Amazon Resource Name (ARN) of the IAM role that allows Amazon SNS to record Firewall Manager activity.

Destructuring the Response

data PutNotificationChannelResponse

See: newPutNotificationChannelResponse smart constructor.

Constructors

PutNotificationChannelResponse'

newPutNotificationChannelResponse :: PutNotificationChannelResponse

Create a value of PutNotificationChannelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.