Network.AWS.FMS.Types.AppsListData

Description

Documentation

data AppsListData

An Firewall Manager applications list.

See: newAppsListData smart constructor.

Constructors

AppsListData'

Fields

lastUpdateTime :: Maybe POSIX

The time that the Firewall Manager applications list was last updated.

listId :: Maybe Text

The ID of the Firewall Manager applications list.

createTime :: Maybe POSIX

The time that the Firewall Manager applications list was created.

previousAppsList :: Maybe (HashMap Text [App])

A map of previous version numbers to their corresponding App object arrays.

listUpdateToken :: Maybe Text

A unique identifier for each update to the list. When you update the list, the update token must match the token of the current version of the application list. You can retrieve the update token by getting the list.

listName :: Text

The name of the Firewall Manager applications list.

appsList :: [App]

An array of applications in the Firewall Manager applications list.

newAppsListData

Arguments

:: Text
$sel:listName:AppsListData'

-> AppsListData

Create a value of AppsListData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastUpdateTime:AppsListData', appsListData_lastUpdateTime - The time that the Firewall Manager applications list was last updated.

$sel:listId:AppsListData', appsListData_listId - The ID of the Firewall Manager applications list.

$sel:createTime:AppsListData', appsListData_createTime - The time that the Firewall Manager applications list was created.

$sel:previousAppsList:AppsListData', appsListData_previousAppsList - A map of previous version numbers to their corresponding App object arrays.

$sel:listUpdateToken:AppsListData', appsListData_listUpdateToken - A unique identifier for each update to the list. When you update the list, the update token must match the token of the current version of the application list. You can retrieve the update token by getting the list.

$sel:listName:AppsListData', appsListData_listName - The name of the Firewall Manager applications list.

$sel:appsList:AppsListData', appsListData_appsList - An array of applications in the Firewall Manager applications list.

appsListData_lastUpdateTime :: Lens' AppsListData (Maybe UTCTime)

The time that the Firewall Manager applications list was last updated.

appsListData_listId :: Lens' AppsListData (Maybe Text)

The ID of the Firewall Manager applications list.

appsListData_createTime :: Lens' AppsListData (Maybe UTCTime)

The time that the Firewall Manager applications list was created.

appsListData_previousAppsList :: Lens' AppsListData (Maybe (HashMap Text [App]))

A map of previous version numbers to their corresponding App object arrays.

appsListData_listUpdateToken :: Lens' AppsListData (Maybe Text)

A unique identifier for each update to the list. When you update the list, the update token must match the token of the current version of the application list. You can retrieve the update token by getting the list.

appsListData_listName :: Lens' AppsListData Text

The name of the Firewall Manager applications list.

appsListData_appsList :: Lens' AppsListData [App]

An array of applications in the Firewall Manager applications list.