Network.AWS.FMS.Types.AppsListDataSummary

Description

Documentation

data AppsListDataSummary

Details of the Firewall Manager applications list.

See: newAppsListDataSummary smart constructor.

Constructors

AppsListDataSummary'

Fields

appsList :: Maybe [App]

An array of App objects in the Firewall Manager applications list.

listName :: Maybe Text

The name of the applications list.

listId :: Maybe Text

The ID of the applications list.

listArn :: Maybe Text

The Amazon Resource Name (ARN) of the applications list.

newAppsListDataSummary :: AppsListDataSummary

Create a value of AppsListDataSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appsList:AppsListDataSummary', appsListDataSummary_appsList - An array of App objects in the Firewall Manager applications list.

$sel:listName:AppsListDataSummary', appsListDataSummary_listName - The name of the applications list.

$sel:listId:AppsListDataSummary', appsListDataSummary_listId - The ID of the applications list.

$sel:listArn:AppsListDataSummary', appsListDataSummary_listArn - The Amazon Resource Name (ARN) of the applications list.

appsListDataSummary_appsList :: Lens' AppsListDataSummary (Maybe [App])

An array of App objects in the Firewall Manager applications list.

appsListDataSummary_listName :: Lens' AppsListDataSummary (Maybe Text)

The name of the applications list.

appsListDataSummary_listId :: Lens' AppsListDataSummary (Maybe Text)

The ID of the applications list.

appsListDataSummary_listArn :: Lens' AppsListDataSummary (Maybe Text)

The Amazon Resource Name (ARN) of the applications list.