Network.AWS.FMS.Types.AwsEc2NetworkInterfaceViolation

Description

Documentation

data AwsEc2NetworkInterfaceViolation

Violation detail for network interfaces associated with an EC2 instance.

See: newAwsEc2NetworkInterfaceViolation smart constructor.

Constructors

AwsEc2NetworkInterfaceViolation'

Fields

violatingSecurityGroups :: Maybe [Text]

List of security groups that violate the rules specified in the primary security group of the Firewall Manager policy.

violationTarget :: Maybe Text

The resource ID of the network interface.

newAwsEc2NetworkInterfaceViolation :: AwsEc2NetworkInterfaceViolation

Create a value of AwsEc2NetworkInterfaceViolation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:violatingSecurityGroups:AwsEc2NetworkInterfaceViolation', awsEc2NetworkInterfaceViolation_violatingSecurityGroups - List of security groups that violate the rules specified in the primary security group of the Firewall Manager policy.

$sel:violationTarget:AwsEc2NetworkInterfaceViolation', awsEc2NetworkInterfaceViolation_violationTarget - The resource ID of the network interface.

awsEc2NetworkInterfaceViolation_violatingSecurityGroups :: Lens' AwsEc2NetworkInterfaceViolation (Maybe [Text])

List of security groups that violate the rules specified in the primary security group of the Firewall Manager policy.

awsEc2NetworkInterfaceViolation_violationTarget :: Lens' AwsEc2NetworkInterfaceViolation (Maybe Text)

The resource ID of the network interface.