Network.AWS.FMS.Types.AwsVPCSecurityGroupViolation

Description

Documentation

data AwsVPCSecurityGroupViolation

Violation detail for the rule violation in a security group when compared to the primary security group of the Firewall Manager policy.

See: newAwsVPCSecurityGroupViolation smart constructor.

Constructors

AwsVPCSecurityGroupViolation'

Fields

possibleSecurityGroupRemediationActions :: Maybe [SecurityGroupRemediationAction]

Remediation options for the rule specified in the ViolationTarget.

partialMatches :: Maybe [PartialMatch]

List of rules specified in the security group of the Firewall Manager policy that partially match the ViolationTarget rule.

violationTarget :: Maybe Text

The security group rule that is being evaluated.

violationTargetDescription :: Maybe Text

A description of the security group that violates the policy.

newAwsVPCSecurityGroupViolation :: AwsVPCSecurityGroupViolation

Create a value of AwsVPCSecurityGroupViolation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:possibleSecurityGroupRemediationActions:AwsVPCSecurityGroupViolation', awsVPCSecurityGroupViolation_possibleSecurityGroupRemediationActions - Remediation options for the rule specified in the ViolationTarget.

$sel:partialMatches:AwsVPCSecurityGroupViolation', awsVPCSecurityGroupViolation_partialMatches - List of rules specified in the security group of the Firewall Manager policy that partially match the ViolationTarget rule.

$sel:violationTarget:AwsVPCSecurityGroupViolation', awsVPCSecurityGroupViolation_violationTarget - The security group rule that is being evaluated.

$sel:violationTargetDescription:AwsVPCSecurityGroupViolation', awsVPCSecurityGroupViolation_violationTargetDescription - A description of the security group that violates the policy.

awsVPCSecurityGroupViolation_possibleSecurityGroupRemediationActions :: Lens' AwsVPCSecurityGroupViolation (Maybe [SecurityGroupRemediationAction])

Remediation options for the rule specified in the ViolationTarget.

awsVPCSecurityGroupViolation_partialMatches :: Lens' AwsVPCSecurityGroupViolation (Maybe [PartialMatch])

List of rules specified in the security group of the Firewall Manager policy that partially match the ViolationTarget rule.

awsVPCSecurityGroupViolation_violationTarget :: Lens' AwsVPCSecurityGroupViolation (Maybe Text)

The security group rule that is being evaluated.

awsVPCSecurityGroupViolation_violationTargetDescription :: Lens' AwsVPCSecurityGroupViolation (Maybe Text)

A description of the security group that violates the policy.