Network.AWS.FMS.Types.ExpectedRoute

Description

Documentation

data ExpectedRoute

Information about the expected route in the route table.

See: newExpectedRoute smart constructor.

Constructors

ExpectedRoute'

Fields

ipV4Cidr :: Maybe Text

Information about the IPv4 CIDR block.

routeTableId :: Maybe Text

Information about the route table ID.

allowedTargets :: Maybe [Text]

Information about the allowed targets.

prefixListId :: Maybe Text

Information about the ID of the prefix list for the route.

ipV6Cidr :: Maybe Text

Information about the IPv6 CIDR block.

contributingSubnets :: Maybe [Text]

Information about the contributing subnets.

newExpectedRoute :: ExpectedRoute

Create a value of ExpectedRoute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ipV4Cidr:ExpectedRoute', expectedRoute_ipV4Cidr - Information about the IPv4 CIDR block.

$sel:routeTableId:ExpectedRoute', expectedRoute_routeTableId - Information about the route table ID.

$sel:allowedTargets:ExpectedRoute', expectedRoute_allowedTargets - Information about the allowed targets.

$sel:prefixListId:ExpectedRoute', expectedRoute_prefixListId - Information about the ID of the prefix list for the route.

$sel:ipV6Cidr:ExpectedRoute', expectedRoute_ipV6Cidr - Information about the IPv6 CIDR block.

$sel:contributingSubnets:ExpectedRoute', expectedRoute_contributingSubnets - Information about the contributing subnets.

expectedRoute_ipV4Cidr :: Lens' ExpectedRoute (Maybe Text)

Information about the IPv4 CIDR block.

expectedRoute_routeTableId :: Lens' ExpectedRoute (Maybe Text)

Information about the route table ID.

expectedRoute_allowedTargets :: Lens' ExpectedRoute (Maybe [Text])

Information about the allowed targets.

expectedRoute_prefixListId :: Lens' ExpectedRoute (Maybe Text)

Information about the ID of the prefix list for the route.

expectedRoute_ipV6Cidr :: Lens' ExpectedRoute (Maybe Text)

Information about the IPv6 CIDR block.

expectedRoute_contributingSubnets :: Lens' ExpectedRoute (Maybe [Text])

Information about the contributing subnets.