Network.AWS.FMS.Types.NetworkFirewallInternetTrafficNotInspectedViolation

Description

Documentation

data NetworkFirewallInternetTrafficNotInspectedViolation

Violation detail for the subnet for which internet traffic that hasn't been inspected.

See: newNetworkFirewallInternetTrafficNotInspectedViolation smart constructor.

Constructors

NetworkFirewallInternetTrafficNotInspectedViolation'

Fields

actualInternetGatewayRoutes :: Maybe [Route]

The actual internet gateway routes.

routeTableId :: Maybe Text

Information about the route table ID.

vpcId :: Maybe Text

Information about the VPC ID.

violatingRoutes :: Maybe [Route]

The route or routes that are in violation.

subnetId :: Maybe Text

The subnet ID.

expectedFirewallSubnetRoutes :: Maybe [ExpectedRoute]

The firewall subnet routes that are expected.

subnetAvailabilityZone :: Maybe Text

The subnet Availability Zone.

expectedInternetGatewayRoutes :: Maybe [ExpectedRoute]

The internet gateway routes that are expected.

currentInternetGatewayRouteTable :: Maybe Text

The current route table for the internet gateway.

firewallSubnetId :: Maybe Text

The firewall subnet ID.

currentFirewallSubnetRouteTable :: Maybe Text

Information about the subnet route table for the current firewall.

expectedFirewallEndpoint :: Maybe Text

The expected endpoint for the current firewall.

isRouteTableUsedInDifferentAZ :: Maybe Bool

Information about whether the route table is used in another Availability Zone.

internetGatewayId :: Maybe Text

The internet gateway ID.

actualFirewallSubnetRoutes :: Maybe [Route]

The actual firewall subnet routes.

newNetworkFirewallInternetTrafficNotInspectedViolation :: NetworkFirewallInternetTrafficNotInspectedViolation

Create a value of NetworkFirewallInternetTrafficNotInspectedViolation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:actualInternetGatewayRoutes:NetworkFirewallInternetTrafficNotInspectedViolation', networkFirewallInternetTrafficNotInspectedViolation_actualInternetGatewayRoutes - The actual internet gateway routes.

$sel:routeTableId:NetworkFirewallInternetTrafficNotInspectedViolation', networkFirewallInternetTrafficNotInspectedViolation_routeTableId - Information about the route table ID.

$sel:vpcId:NetworkFirewallInternetTrafficNotInspectedViolation', networkFirewallInternetTrafficNotInspectedViolation_vpcId - Information about the VPC ID.

$sel:violatingRoutes:NetworkFirewallInternetTrafficNotInspectedViolation', networkFirewallInternetTrafficNotInspectedViolation_violatingRoutes - The route or routes that are in violation.

$sel:subnetId:NetworkFirewallInternetTrafficNotInspectedViolation', networkFirewallInternetTrafficNotInspectedViolation_subnetId - The subnet ID.

$sel:expectedFirewallSubnetRoutes:NetworkFirewallInternetTrafficNotInspectedViolation', networkFirewallInternetTrafficNotInspectedViolation_expectedFirewallSubnetRoutes - The firewall subnet routes that are expected.

$sel:subnetAvailabilityZone:NetworkFirewallInternetTrafficNotInspectedViolation', networkFirewallInternetTrafficNotInspectedViolation_subnetAvailabilityZone - The subnet Availability Zone.

$sel:expectedInternetGatewayRoutes:NetworkFirewallInternetTrafficNotInspectedViolation', networkFirewallInternetTrafficNotInspectedViolation_expectedInternetGatewayRoutes - The internet gateway routes that are expected.

$sel:currentInternetGatewayRouteTable:NetworkFirewallInternetTrafficNotInspectedViolation', networkFirewallInternetTrafficNotInspectedViolation_currentInternetGatewayRouteTable - The current route table for the internet gateway.

$sel:firewallSubnetId:NetworkFirewallInternetTrafficNotInspectedViolation', networkFirewallInternetTrafficNotInspectedViolation_firewallSubnetId - The firewall subnet ID.

$sel:currentFirewallSubnetRouteTable:NetworkFirewallInternetTrafficNotInspectedViolation', networkFirewallInternetTrafficNotInspectedViolation_currentFirewallSubnetRouteTable - Information about the subnet route table for the current firewall.

$sel:expectedFirewallEndpoint:NetworkFirewallInternetTrafficNotInspectedViolation', networkFirewallInternetTrafficNotInspectedViolation_expectedFirewallEndpoint - The expected endpoint for the current firewall.

$sel:isRouteTableUsedInDifferentAZ:NetworkFirewallInternetTrafficNotInspectedViolation', networkFirewallInternetTrafficNotInspectedViolation_isRouteTableUsedInDifferentAZ - Information about whether the route table is used in another Availability Zone.

$sel:internetGatewayId:NetworkFirewallInternetTrafficNotInspectedViolation', networkFirewallInternetTrafficNotInspectedViolation_internetGatewayId - The internet gateway ID.

$sel:actualFirewallSubnetRoutes:NetworkFirewallInternetTrafficNotInspectedViolation', networkFirewallInternetTrafficNotInspectedViolation_actualFirewallSubnetRoutes - The actual firewall subnet routes.

networkFirewallInternetTrafficNotInspectedViolation_actualInternetGatewayRoutes :: Lens' NetworkFirewallInternetTrafficNotInspectedViolation (Maybe [Route])

The actual internet gateway routes.

networkFirewallInternetTrafficNotInspectedViolation_routeTableId :: Lens' NetworkFirewallInternetTrafficNotInspectedViolation (Maybe Text)

Information about the route table ID.

networkFirewallInternetTrafficNotInspectedViolation_vpcId :: Lens' NetworkFirewallInternetTrafficNotInspectedViolation (Maybe Text)

Information about the VPC ID.

networkFirewallInternetTrafficNotInspectedViolation_violatingRoutes :: Lens' NetworkFirewallInternetTrafficNotInspectedViolation (Maybe [Route])

The route or routes that are in violation.

networkFirewallInternetTrafficNotInspectedViolation_subnetId :: Lens' NetworkFirewallInternetTrafficNotInspectedViolation (Maybe Text)

The subnet ID.

networkFirewallInternetTrafficNotInspectedViolation_expectedFirewallSubnetRoutes :: Lens' NetworkFirewallInternetTrafficNotInspectedViolation (Maybe [ExpectedRoute])

The firewall subnet routes that are expected.

networkFirewallInternetTrafficNotInspectedViolation_subnetAvailabilityZone :: Lens' NetworkFirewallInternetTrafficNotInspectedViolation (Maybe Text)

The subnet Availability Zone.

networkFirewallInternetTrafficNotInspectedViolation_expectedInternetGatewayRoutes :: Lens' NetworkFirewallInternetTrafficNotInspectedViolation (Maybe [ExpectedRoute])

The internet gateway routes that are expected.

networkFirewallInternetTrafficNotInspectedViolation_currentInternetGatewayRouteTable :: Lens' NetworkFirewallInternetTrafficNotInspectedViolation (Maybe Text)

The current route table for the internet gateway.

networkFirewallInternetTrafficNotInspectedViolation_firewallSubnetId :: Lens' NetworkFirewallInternetTrafficNotInspectedViolation (Maybe Text)

The firewall subnet ID.

networkFirewallInternetTrafficNotInspectedViolation_currentFirewallSubnetRouteTable :: Lens' NetworkFirewallInternetTrafficNotInspectedViolation (Maybe Text)

Information about the subnet route table for the current firewall.

networkFirewallInternetTrafficNotInspectedViolation_expectedFirewallEndpoint :: Lens' NetworkFirewallInternetTrafficNotInspectedViolation (Maybe Text)

The expected endpoint for the current firewall.

networkFirewallInternetTrafficNotInspectedViolation_isRouteTableUsedInDifferentAZ :: Lens' NetworkFirewallInternetTrafficNotInspectedViolation (Maybe Bool)

Information about whether the route table is used in another Availability Zone.

networkFirewallInternetTrafficNotInspectedViolation_internetGatewayId :: Lens' NetworkFirewallInternetTrafficNotInspectedViolation (Maybe Text)

The internet gateway ID.

networkFirewallInternetTrafficNotInspectedViolation_actualFirewallSubnetRoutes :: Lens' NetworkFirewallInternetTrafficNotInspectedViolation (Maybe [Route])

The actual firewall subnet routes.