Network.AWS.FMS.Types.NetworkFirewallInvalidRouteConfigurationViolation

Description

Documentation

data NetworkFirewallInvalidRouteConfigurationViolation

Violation detail for the improperly configured subnet route. It's possible there is a missing route table route, or a configuration that causes traffic to cross an Availability Zone boundary.

See: newNetworkFirewallInvalidRouteConfigurationViolation smart constructor.

Constructors

NetworkFirewallInvalidRouteConfigurationViolation'

Fields

actualInternetGatewayRoutes :: Maybe [Route]

The actual internet gateway routes.

routeTableId :: Maybe Text

The route table ID.

affectedSubnets :: Maybe [Text]

The subnets that are affected.

vpcId :: Maybe Text

Information about the VPC ID.

actualFirewallEndpoint :: Maybe Text

The actual firewall endpoint.

expectedFirewallSubnetId :: Maybe Text

The expected subnet ID for the firewall.

expectedFirewallSubnetRoutes :: Maybe [ExpectedRoute]

The firewall subnet routes that are expected.

expectedInternetGatewayRoutes :: Maybe [ExpectedRoute]

The expected routes for the internet gateway.

currentInternetGatewayRouteTable :: Maybe Text

The route table for the current internet gateway.

violatingRoute :: Maybe Route

The route that's in violation.

currentFirewallSubnetRouteTable :: Maybe Text

The subnet route table for the current firewall.

expectedFirewallEndpoint :: Maybe Text

The firewall endpoint that's expected.

isRouteTableUsedInDifferentAZ :: Maybe Bool

Information about whether the route table is used in another Availability Zone.

actualFirewallSubnetId :: Maybe Text

The actual subnet ID for the firewall.

internetGatewayId :: Maybe Text

The internet gateway ID.

actualFirewallSubnetRoutes :: Maybe [Route]

The actual firewall subnet routes that are expected.

newNetworkFirewallInvalidRouteConfigurationViolation :: NetworkFirewallInvalidRouteConfigurationViolation

Create a value of NetworkFirewallInvalidRouteConfigurationViolation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:actualInternetGatewayRoutes:NetworkFirewallInvalidRouteConfigurationViolation', networkFirewallInvalidRouteConfigurationViolation_actualInternetGatewayRoutes - The actual internet gateway routes.

$sel:routeTableId:NetworkFirewallInvalidRouteConfigurationViolation', networkFirewallInvalidRouteConfigurationViolation_routeTableId - The route table ID.

$sel:affectedSubnets:NetworkFirewallInvalidRouteConfigurationViolation', networkFirewallInvalidRouteConfigurationViolation_affectedSubnets - The subnets that are affected.

$sel:vpcId:NetworkFirewallInvalidRouteConfigurationViolation', networkFirewallInvalidRouteConfigurationViolation_vpcId - Information about the VPC ID.

$sel:actualFirewallEndpoint:NetworkFirewallInvalidRouteConfigurationViolation', networkFirewallInvalidRouteConfigurationViolation_actualFirewallEndpoint - The actual firewall endpoint.

$sel:expectedFirewallSubnetId:NetworkFirewallInvalidRouteConfigurationViolation', networkFirewallInvalidRouteConfigurationViolation_expectedFirewallSubnetId - The expected subnet ID for the firewall.

$sel:expectedFirewallSubnetRoutes:NetworkFirewallInvalidRouteConfigurationViolation', networkFirewallInvalidRouteConfigurationViolation_expectedFirewallSubnetRoutes - The firewall subnet routes that are expected.

$sel:expectedInternetGatewayRoutes:NetworkFirewallInvalidRouteConfigurationViolation', networkFirewallInvalidRouteConfigurationViolation_expectedInternetGatewayRoutes - The expected routes for the internet gateway.

$sel:currentInternetGatewayRouteTable:NetworkFirewallInvalidRouteConfigurationViolation', networkFirewallInvalidRouteConfigurationViolation_currentInternetGatewayRouteTable - The route table for the current internet gateway.

$sel:violatingRoute:NetworkFirewallInvalidRouteConfigurationViolation', networkFirewallInvalidRouteConfigurationViolation_violatingRoute - The route that's in violation.

$sel:currentFirewallSubnetRouteTable:NetworkFirewallInvalidRouteConfigurationViolation', networkFirewallInvalidRouteConfigurationViolation_currentFirewallSubnetRouteTable - The subnet route table for the current firewall.

$sel:expectedFirewallEndpoint:NetworkFirewallInvalidRouteConfigurationViolation', networkFirewallInvalidRouteConfigurationViolation_expectedFirewallEndpoint - The firewall endpoint that's expected.

$sel:isRouteTableUsedInDifferentAZ:NetworkFirewallInvalidRouteConfigurationViolation', networkFirewallInvalidRouteConfigurationViolation_isRouteTableUsedInDifferentAZ - Information about whether the route table is used in another Availability Zone.

$sel:actualFirewallSubnetId:NetworkFirewallInvalidRouteConfigurationViolation', networkFirewallInvalidRouteConfigurationViolation_actualFirewallSubnetId - The actual subnet ID for the firewall.

$sel:internetGatewayId:NetworkFirewallInvalidRouteConfigurationViolation', networkFirewallInvalidRouteConfigurationViolation_internetGatewayId - The internet gateway ID.

$sel:actualFirewallSubnetRoutes:NetworkFirewallInvalidRouteConfigurationViolation', networkFirewallInvalidRouteConfigurationViolation_actualFirewallSubnetRoutes - The actual firewall subnet routes that are expected.

networkFirewallInvalidRouteConfigurationViolation_actualInternetGatewayRoutes :: Lens' NetworkFirewallInvalidRouteConfigurationViolation (Maybe [Route])

The actual internet gateway routes.

networkFirewallInvalidRouteConfigurationViolation_routeTableId :: Lens' NetworkFirewallInvalidRouteConfigurationViolation (Maybe Text)

The route table ID.

networkFirewallInvalidRouteConfigurationViolation_affectedSubnets :: Lens' NetworkFirewallInvalidRouteConfigurationViolation (Maybe [Text])

The subnets that are affected.

networkFirewallInvalidRouteConfigurationViolation_vpcId :: Lens' NetworkFirewallInvalidRouteConfigurationViolation (Maybe Text)

Information about the VPC ID.

networkFirewallInvalidRouteConfigurationViolation_actualFirewallEndpoint :: Lens' NetworkFirewallInvalidRouteConfigurationViolation (Maybe Text)

The actual firewall endpoint.

networkFirewallInvalidRouteConfigurationViolation_expectedFirewallSubnetId :: Lens' NetworkFirewallInvalidRouteConfigurationViolation (Maybe Text)

The expected subnet ID for the firewall.

networkFirewallInvalidRouteConfigurationViolation_expectedFirewallSubnetRoutes :: Lens' NetworkFirewallInvalidRouteConfigurationViolation (Maybe [ExpectedRoute])

The firewall subnet routes that are expected.

networkFirewallInvalidRouteConfigurationViolation_expectedInternetGatewayRoutes :: Lens' NetworkFirewallInvalidRouteConfigurationViolation (Maybe [ExpectedRoute])

The expected routes for the internet gateway.

networkFirewallInvalidRouteConfigurationViolation_currentInternetGatewayRouteTable :: Lens' NetworkFirewallInvalidRouteConfigurationViolation (Maybe Text)

The route table for the current internet gateway.

networkFirewallInvalidRouteConfigurationViolation_violatingRoute :: Lens' NetworkFirewallInvalidRouteConfigurationViolation (Maybe Route)

The route that's in violation.

networkFirewallInvalidRouteConfigurationViolation_currentFirewallSubnetRouteTable :: Lens' NetworkFirewallInvalidRouteConfigurationViolation (Maybe Text)

The subnet route table for the current firewall.

networkFirewallInvalidRouteConfigurationViolation_expectedFirewallEndpoint :: Lens' NetworkFirewallInvalidRouteConfigurationViolation (Maybe Text)

The firewall endpoint that's expected.

networkFirewallInvalidRouteConfigurationViolation_isRouteTableUsedInDifferentAZ :: Lens' NetworkFirewallInvalidRouteConfigurationViolation (Maybe Bool)

Information about whether the route table is used in another Availability Zone.

networkFirewallInvalidRouteConfigurationViolation_actualFirewallSubnetId :: Lens' NetworkFirewallInvalidRouteConfigurationViolation (Maybe Text)

The actual subnet ID for the firewall.

networkFirewallInvalidRouteConfigurationViolation_internetGatewayId :: Lens' NetworkFirewallInvalidRouteConfigurationViolation (Maybe Text)

The internet gateway ID.

networkFirewallInvalidRouteConfigurationViolation_actualFirewallSubnetRoutes :: Lens' NetworkFirewallInvalidRouteConfigurationViolation (Maybe [Route])

The actual firewall subnet routes that are expected.