Network.AWS.FMS.Types.NetworkFirewallPolicyDescription

Description

Documentation

data NetworkFirewallPolicyDescription

The definition of the Network Firewall firewall policy.

See: newNetworkFirewallPolicyDescription smart constructor.

Constructors

NetworkFirewallPolicyDescription'

Fields

statelessRuleGroups :: Maybe [StatelessRuleGroup]

The stateless rule groups that are used in the Network Firewall firewall policy.

statelessDefaultActions :: Maybe [Text]

The actions to take on packets that don't match any of the stateless rule groups.

statelessFragmentDefaultActions :: Maybe [Text]

The actions to take on packet fragments that don't match any of the stateless rule groups.

statefulRuleGroups :: Maybe [StatefulRuleGroup]

The stateful rule groups that are used in the Network Firewall firewall policy.

statelessCustomActions :: Maybe [Text]

Names of custom actions that are available for use in the stateless default actions settings.

newNetworkFirewallPolicyDescription :: NetworkFirewallPolicyDescription

Create a value of NetworkFirewallPolicyDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:statelessRuleGroups:NetworkFirewallPolicyDescription', networkFirewallPolicyDescription_statelessRuleGroups - The stateless rule groups that are used in the Network Firewall firewall policy.

$sel:statelessDefaultActions:NetworkFirewallPolicyDescription', networkFirewallPolicyDescription_statelessDefaultActions - The actions to take on packets that don't match any of the stateless rule groups.

$sel:statelessFragmentDefaultActions:NetworkFirewallPolicyDescription', networkFirewallPolicyDescription_statelessFragmentDefaultActions - The actions to take on packet fragments that don't match any of the stateless rule groups.

$sel:statefulRuleGroups:NetworkFirewallPolicyDescription', networkFirewallPolicyDescription_statefulRuleGroups - The stateful rule groups that are used in the Network Firewall firewall policy.

$sel:statelessCustomActions:NetworkFirewallPolicyDescription', networkFirewallPolicyDescription_statelessCustomActions - Names of custom actions that are available for use in the stateless default actions settings.

networkFirewallPolicyDescription_statelessRuleGroups :: Lens' NetworkFirewallPolicyDescription (Maybe [StatelessRuleGroup])

The stateless rule groups that are used in the Network Firewall firewall policy.

networkFirewallPolicyDescription_statelessDefaultActions :: Lens' NetworkFirewallPolicyDescription (Maybe [Text])

The actions to take on packets that don't match any of the stateless rule groups.

networkFirewallPolicyDescription_statelessFragmentDefaultActions :: Lens' NetworkFirewallPolicyDescription (Maybe [Text])

The actions to take on packet fragments that don't match any of the stateless rule groups.

networkFirewallPolicyDescription_statefulRuleGroups :: Lens' NetworkFirewallPolicyDescription (Maybe [StatefulRuleGroup])

The stateful rule groups that are used in the Network Firewall firewall policy.

networkFirewallPolicyDescription_statelessCustomActions :: Lens' NetworkFirewallPolicyDescription (Maybe [Text])

Names of custom actions that are available for use in the stateless default actions settings.