Network.AWS.FMS.Types.PolicyComplianceDetail

Description

Documentation

data PolicyComplianceDetail

Describes the noncompliant resources in a member account for a specific Firewall Manager policy. A maximum of 100 entries are displayed. If more than 100 resources are noncompliant, EvaluationLimitExceeded is set to True.

See: newPolicyComplianceDetail smart constructor.

Constructors

PolicyComplianceDetail'

Fields

violators :: Maybe [ComplianceViolator]

An array of resources that aren't protected by the WAF or Shield Advanced policy or that aren't in compliance with the security group policy.

policyOwner :: Maybe Text

The Amazon Web Services account that created the Firewall Manager policy.

memberAccount :: Maybe Text

The Amazon Web Services account ID.

evaluationLimitExceeded :: Maybe Bool

Indicates if over 100 resources are noncompliant with the Firewall Manager policy.

issueInfoMap :: Maybe (HashMap DependentServiceName Text)

Details about problems with dependent services, such as WAF or Config, and the error message received that indicates the problem with the service.

policyId :: Maybe Text

The ID of the Firewall Manager policy.

expiredAt :: Maybe POSIX

A timestamp that indicates when the returned information should be considered out of date.

newPolicyComplianceDetail :: PolicyComplianceDetail

Create a value of PolicyComplianceDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:violators:PolicyComplianceDetail', policyComplianceDetail_violators - An array of resources that aren't protected by the WAF or Shield Advanced policy or that aren't in compliance with the security group policy.

$sel:policyOwner:PolicyComplianceDetail', policyComplianceDetail_policyOwner - The Amazon Web Services account that created the Firewall Manager policy.

$sel:memberAccount:PolicyComplianceDetail', policyComplianceDetail_memberAccount - The Amazon Web Services account ID.

$sel:evaluationLimitExceeded:PolicyComplianceDetail', policyComplianceDetail_evaluationLimitExceeded - Indicates if over 100 resources are noncompliant with the Firewall Manager policy.

$sel:issueInfoMap:PolicyComplianceDetail', policyComplianceDetail_issueInfoMap - Details about problems with dependent services, such as WAF or Config, and the error message received that indicates the problem with the service.

$sel:policyId:PolicyComplianceDetail', policyComplianceDetail_policyId - The ID of the Firewall Manager policy.

$sel:expiredAt:PolicyComplianceDetail', policyComplianceDetail_expiredAt - A timestamp that indicates when the returned information should be considered out of date.

policyComplianceDetail_violators :: Lens' PolicyComplianceDetail (Maybe [ComplianceViolator])

An array of resources that aren't protected by the WAF or Shield Advanced policy or that aren't in compliance with the security group policy.

policyComplianceDetail_policyOwner :: Lens' PolicyComplianceDetail (Maybe Text)

The Amazon Web Services account that created the Firewall Manager policy.

policyComplianceDetail_memberAccount :: Lens' PolicyComplianceDetail (Maybe Text)

The Amazon Web Services account ID.

policyComplianceDetail_evaluationLimitExceeded :: Lens' PolicyComplianceDetail (Maybe Bool)

Indicates if over 100 resources are noncompliant with the Firewall Manager policy.

policyComplianceDetail_issueInfoMap :: Lens' PolicyComplianceDetail (Maybe (HashMap DependentServiceName Text))

Details about problems with dependent services, such as WAF or Config, and the error message received that indicates the problem with the service.

policyComplianceDetail_policyId :: Lens' PolicyComplianceDetail (Maybe Text)

The ID of the Firewall Manager policy.

policyComplianceDetail_expiredAt :: Lens' PolicyComplianceDetail (Maybe UTCTime)

A timestamp that indicates when the returned information should be considered out of date.