Network.AWS.FMS.Types.PolicyComplianceStatus

Description

Documentation

data PolicyComplianceStatus

Indicates whether the account is compliant with the specified policy. An account is considered noncompliant if it includes resources that are not protected by the policy, for WAF and Shield Advanced policies, or that are noncompliant with the policy, for security group policies.

See: newPolicyComplianceStatus smart constructor.

Constructors

PolicyComplianceStatus'

Fields

policyName :: Maybe Text

The name of the Firewall Manager policy.

policyOwner :: Maybe Text

The Amazon Web Services account that created the Firewall Manager policy.

memberAccount :: Maybe Text

The member account ID.

lastUpdated :: Maybe POSIX

Timestamp of the last update to the EvaluationResult objects.

evaluationResults :: Maybe [EvaluationResult]

An array of EvaluationResult objects.

issueInfoMap :: Maybe (HashMap DependentServiceName Text)

Details about problems with dependent services, such as WAF or Config, and the error message received that indicates the problem with the service.

policyId :: Maybe Text

The ID of the Firewall Manager policy.

newPolicyComplianceStatus :: PolicyComplianceStatus

Create a value of PolicyComplianceStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policyName:PolicyComplianceStatus', policyComplianceStatus_policyName - The name of the Firewall Manager policy.

$sel:policyOwner:PolicyComplianceStatus', policyComplianceStatus_policyOwner - The Amazon Web Services account that created the Firewall Manager policy.

$sel:memberAccount:PolicyComplianceStatus', policyComplianceStatus_memberAccount - The member account ID.

$sel:lastUpdated:PolicyComplianceStatus', policyComplianceStatus_lastUpdated - Timestamp of the last update to the EvaluationResult objects.

$sel:evaluationResults:PolicyComplianceStatus', policyComplianceStatus_evaluationResults - An array of EvaluationResult objects.

$sel:issueInfoMap:PolicyComplianceStatus', policyComplianceStatus_issueInfoMap - Details about problems with dependent services, such as WAF or Config, and the error message received that indicates the problem with the service.

$sel:policyId:PolicyComplianceStatus', policyComplianceStatus_policyId - The ID of the Firewall Manager policy.

policyComplianceStatus_policyName :: Lens' PolicyComplianceStatus (Maybe Text)

The name of the Firewall Manager policy.

policyComplianceStatus_policyOwner :: Lens' PolicyComplianceStatus (Maybe Text)

The Amazon Web Services account that created the Firewall Manager policy.

policyComplianceStatus_memberAccount :: Lens' PolicyComplianceStatus (Maybe Text)

The member account ID.

policyComplianceStatus_lastUpdated :: Lens' PolicyComplianceStatus (Maybe UTCTime)

Timestamp of the last update to the EvaluationResult objects.

policyComplianceStatus_evaluationResults :: Lens' PolicyComplianceStatus (Maybe [EvaluationResult])

An array of EvaluationResult objects.

policyComplianceStatus_issueInfoMap :: Lens' PolicyComplianceStatus (Maybe (HashMap DependentServiceName Text))

Details about problems with dependent services, such as WAF or Config, and the error message received that indicates the problem with the service.

policyComplianceStatus_policyId :: Lens' PolicyComplianceStatus (Maybe Text)

The ID of the Firewall Manager policy.