Network.AWS.FMS.Types.ProtocolsListData

Description

Documentation

data ProtocolsListData

An Firewall Manager protocols list.

See: newProtocolsListData smart constructor.

Constructors

ProtocolsListData'

Fields

listUpdateToken :: Maybe Text

A unique identifier for each update to the list. When you update the list, the update token must match the token of the current version of the application list. You can retrieve the update token by getting the list.

listId :: Maybe Text

The ID of the Firewall Manager protocols list.

lastUpdateTime :: Maybe POSIX

The time that the Firewall Manager protocols list was last updated.

previousProtocolsList :: Maybe (HashMap Text [Text])

A map of previous version numbers to their corresponding protocol arrays.

createTime :: Maybe POSIX

The time that the Firewall Manager protocols list was created.

listName :: Text

The name of the Firewall Manager protocols list.

protocolsList :: [Text]

An array of protocols in the Firewall Manager protocols list.

newProtocolsListData

Arguments

:: Text
$sel:listName:ProtocolsListData'

-> ProtocolsListData

Create a value of ProtocolsListData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:listUpdateToken:ProtocolsListData', protocolsListData_listUpdateToken - A unique identifier for each update to the list. When you update the list, the update token must match the token of the current version of the application list. You can retrieve the update token by getting the list.

$sel:listId:ProtocolsListData', protocolsListData_listId - The ID of the Firewall Manager protocols list.

$sel:lastUpdateTime:ProtocolsListData', protocolsListData_lastUpdateTime - The time that the Firewall Manager protocols list was last updated.

$sel:previousProtocolsList:ProtocolsListData', protocolsListData_previousProtocolsList - A map of previous version numbers to their corresponding protocol arrays.

$sel:createTime:ProtocolsListData', protocolsListData_createTime - The time that the Firewall Manager protocols list was created.

$sel:listName:ProtocolsListData', protocolsListData_listName - The name of the Firewall Manager protocols list.

$sel:protocolsList:ProtocolsListData', protocolsListData_protocolsList - An array of protocols in the Firewall Manager protocols list.

protocolsListData_listUpdateToken :: Lens' ProtocolsListData (Maybe Text)

A unique identifier for each update to the list. When you update the list, the update token must match the token of the current version of the application list. You can retrieve the update token by getting the list.

protocolsListData_listId :: Lens' ProtocolsListData (Maybe Text)

The ID of the Firewall Manager protocols list.

protocolsListData_lastUpdateTime :: Lens' ProtocolsListData (Maybe UTCTime)

The time that the Firewall Manager protocols list was last updated.

protocolsListData_previousProtocolsList :: Lens' ProtocolsListData (Maybe (HashMap Text [Text]))

A map of previous version numbers to their corresponding protocol arrays.

protocolsListData_createTime :: Lens' ProtocolsListData (Maybe UTCTime)

The time that the Firewall Manager protocols list was created.

protocolsListData_listName :: Lens' ProtocolsListData Text

The name of the Firewall Manager protocols list.

protocolsListData_protocolsList :: Lens' ProtocolsListData [Text]

An array of protocols in the Firewall Manager protocols list.