Network.AWS.FMS.Types.RemediationAction

Description

Documentation

data RemediationAction

Information about an individual action you can take to remediate a violation.

See: newRemediationAction smart constructor.

Constructors

RemediationAction'

Fields

eC2CreateRouteAction :: Maybe EC2CreateRouteAction

Information about the CreateRoute action in the Amazon EC2 API.

eC2ReplaceRouteAction :: Maybe EC2ReplaceRouteAction

Information about the ReplaceRoute action in the Amazon EC2 API.

eC2AssociateRouteTableAction :: Maybe EC2AssociateRouteTableAction

Information about the AssociateRouteTable action in the Amazon EC2 API.

eC2CreateRouteTableAction :: Maybe EC2CreateRouteTableAction

Information about the CreateRouteTable action in the Amazon EC2 API.

description :: Maybe Text

A description of a remediation action.

eC2CopyRouteTableAction :: Maybe EC2CopyRouteTableAction

Information about the CopyRouteTable action in the Amazon EC2 API.

eC2ReplaceRouteTableAssociationAction :: Maybe EC2ReplaceRouteTableAssociationAction

Information about the ReplaceRouteTableAssociation action in the Amazon EC2 API.

eC2DeleteRouteAction :: Maybe EC2DeleteRouteAction

Information about the DeleteRoute action in the Amazon EC2 API.

newRemediationAction :: RemediationAction

Create a value of RemediationAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eC2CreateRouteAction:RemediationAction', remediationAction_eC2CreateRouteAction - Information about the CreateRoute action in the Amazon EC2 API.

$sel:eC2ReplaceRouteAction:RemediationAction', remediationAction_eC2ReplaceRouteAction - Information about the ReplaceRoute action in the Amazon EC2 API.

$sel:eC2AssociateRouteTableAction:RemediationAction', remediationAction_eC2AssociateRouteTableAction - Information about the AssociateRouteTable action in the Amazon EC2 API.

$sel:eC2CreateRouteTableAction:RemediationAction', remediationAction_eC2CreateRouteTableAction - Information about the CreateRouteTable action in the Amazon EC2 API.

$sel:description:RemediationAction', remediationAction_description - A description of a remediation action.

$sel:eC2CopyRouteTableAction:RemediationAction', remediationAction_eC2CopyRouteTableAction - Information about the CopyRouteTable action in the Amazon EC2 API.

$sel:eC2ReplaceRouteTableAssociationAction:RemediationAction', remediationAction_eC2ReplaceRouteTableAssociationAction - Information about the ReplaceRouteTableAssociation action in the Amazon EC2 API.

$sel:eC2DeleteRouteAction:RemediationAction', remediationAction_eC2DeleteRouteAction - Information about the DeleteRoute action in the Amazon EC2 API.

remediationAction_eC2CreateRouteAction :: Lens' RemediationAction (Maybe EC2CreateRouteAction)

Information about the CreateRoute action in the Amazon EC2 API.

remediationAction_eC2ReplaceRouteAction :: Lens' RemediationAction (Maybe EC2ReplaceRouteAction)

Information about the ReplaceRoute action in the Amazon EC2 API.

remediationAction_eC2AssociateRouteTableAction :: Lens' RemediationAction (Maybe EC2AssociateRouteTableAction)

Information about the AssociateRouteTable action in the Amazon EC2 API.

remediationAction_eC2CreateRouteTableAction :: Lens' RemediationAction (Maybe EC2CreateRouteTableAction)

Information about the CreateRouteTable action in the Amazon EC2 API.

remediationAction_description :: Lens' RemediationAction (Maybe Text)

A description of a remediation action.

remediationAction_eC2CopyRouteTableAction :: Lens' RemediationAction (Maybe EC2CopyRouteTableAction)

Information about the CopyRouteTable action in the Amazon EC2 API.

remediationAction_eC2ReplaceRouteTableAssociationAction :: Lens' RemediationAction (Maybe EC2ReplaceRouteTableAssociationAction)

Information about the ReplaceRouteTableAssociation action in the Amazon EC2 API.

remediationAction_eC2DeleteRouteAction :: Lens' RemediationAction (Maybe EC2DeleteRouteAction)

Information about the DeleteRoute action in the Amazon EC2 API.