Network.AWS.FMS.Types.RemediationActionWithOrder

Description

Documentation

data RemediationActionWithOrder

An ordered list of actions you can take to remediate a violation.

See: newRemediationActionWithOrder smart constructor.

Constructors

RemediationActionWithOrder'

Fields

remediationAction :: Maybe RemediationAction

Information about an action you can take to remediate a violation.

order :: Maybe Int

The order of the remediation actions in the list.

newRemediationActionWithOrder :: RemediationActionWithOrder

Create a value of RemediationActionWithOrder with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:remediationAction:RemediationActionWithOrder', remediationActionWithOrder_remediationAction - Information about an action you can take to remediate a violation.

$sel:order:RemediationActionWithOrder', remediationActionWithOrder_order - The order of the remediation actions in the list.

remediationActionWithOrder_remediationAction :: Lens' RemediationActionWithOrder (Maybe RemediationAction)

Information about an action you can take to remediate a violation.

remediationActionWithOrder_order :: Lens' RemediationActionWithOrder (Maybe Int)

The order of the remediation actions in the list.