Network.AWS.FMS.Types.ResourceTag

Description

Documentation

data ResourceTag

The resource tags that Firewall Manager uses to determine if a particular resource should be included or excluded from the Firewall Manager policy. Tags enable you to categorize your Amazon Web Services resources in different ways, for example, by purpose, owner, or environment. Each tag consists of a key and an optional value. Firewall Manager combines the tags with "AND" so that, if you add more than one tag to a policy scope, a resource must have all the specified tags to be included or excluded. For more information, see Working with Tag Editor.

See: newResourceTag smart constructor.

Constructors

ResourceTag'

Fields

value :: Maybe Text

The resource tag value.

key :: Text

The resource tag key.

newResourceTag

Arguments

:: Text
$sel:key:ResourceTag'

-> ResourceTag

Create a value of ResourceTag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:ResourceTag', resourceTag_value - The resource tag value.

$sel:key:ResourceTag', resourceTag_key - The resource tag key.

resourceTag_value :: Lens' ResourceTag (Maybe Text)

The resource tag value.

resourceTag_key :: Lens' ResourceTag Text

The resource tag key.