Network.AWS.FMS.Types.SecurityGroupRemediationAction

Description

Documentation

data SecurityGroupRemediationAction

Remediation option for the rule specified in the ViolationTarget.

See: newSecurityGroupRemediationAction smart constructor.

Constructors

SecurityGroupRemediationAction'

Fields

remediationActionType :: Maybe RemediationActionType

The remediation action that will be performed.

remediationResult :: Maybe SecurityGroupRuleDescription

The final state of the rule specified in the ViolationTarget after it is remediated.

isDefaultAction :: Maybe Bool

Indicates if the current action is the default action.

description :: Maybe Text

Brief description of the action that will be performed.

newSecurityGroupRemediationAction :: SecurityGroupRemediationAction

Create a value of SecurityGroupRemediationAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:remediationActionType:SecurityGroupRemediationAction', securityGroupRemediationAction_remediationActionType - The remediation action that will be performed.

$sel:remediationResult:SecurityGroupRemediationAction', securityGroupRemediationAction_remediationResult - The final state of the rule specified in the ViolationTarget after it is remediated.

$sel:isDefaultAction:SecurityGroupRemediationAction', securityGroupRemediationAction_isDefaultAction - Indicates if the current action is the default action.

$sel:description:SecurityGroupRemediationAction', securityGroupRemediationAction_description - Brief description of the action that will be performed.

securityGroupRemediationAction_remediationActionType :: Lens' SecurityGroupRemediationAction (Maybe RemediationActionType)

The remediation action that will be performed.

securityGroupRemediationAction_remediationResult :: Lens' SecurityGroupRemediationAction (Maybe SecurityGroupRuleDescription)

The final state of the rule specified in the ViolationTarget after it is remediated.

securityGroupRemediationAction_isDefaultAction :: Lens' SecurityGroupRemediationAction (Maybe Bool)

Indicates if the current action is the default action.

securityGroupRemediationAction_description :: Lens' SecurityGroupRemediationAction (Maybe Text)

Brief description of the action that will be performed.