Network.AWS.FMS.Types.SecurityGroupRuleDescription

Description

Documentation

data SecurityGroupRuleDescription

Describes a set of permissions for a security group rule.

See: newSecurityGroupRuleDescription smart constructor.

Constructors

SecurityGroupRuleDescription'

Fields

fromPort :: Maybe Natural

The start of the port range for the TCP and UDP protocols, or an ICMP/ICMPv6 type number. A value of -1 indicates all ICMP/ICMPv6 types.

iPV4Range :: Maybe Text

The IPv4 ranges for the security group rule.

prefixListId :: Maybe Text

The ID of the prefix list for the security group rule.

protocol :: Maybe Text

The IP protocol name (tcp, udp, icmp, icmpv6) or number.

iPV6Range :: Maybe Text

The IPv6 ranges for the security group rule.

toPort :: Maybe Natural

The end of the port range for the TCP and UDP protocols, or an ICMP/ICMPv6 code. A value of -1 indicates all ICMP/ICMPv6 codes.

newSecurityGroupRuleDescription :: SecurityGroupRuleDescription

Create a value of SecurityGroupRuleDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fromPort:SecurityGroupRuleDescription', securityGroupRuleDescription_fromPort - The start of the port range for the TCP and UDP protocols, or an ICMP/ICMPv6 type number. A value of -1 indicates all ICMP/ICMPv6 types.

$sel:iPV4Range:SecurityGroupRuleDescription', securityGroupRuleDescription_iPV4Range - The IPv4 ranges for the security group rule.

$sel:prefixListId:SecurityGroupRuleDescription', securityGroupRuleDescription_prefixListId - The ID of the prefix list for the security group rule.

$sel:protocol:SecurityGroupRuleDescription', securityGroupRuleDescription_protocol - The IP protocol name (tcp, udp, icmp, icmpv6) or number.

$sel:iPV6Range:SecurityGroupRuleDescription', securityGroupRuleDescription_iPV6Range - The IPv6 ranges for the security group rule.

$sel:toPort:SecurityGroupRuleDescription', securityGroupRuleDescription_toPort - The end of the port range for the TCP and UDP protocols, or an ICMP/ICMPv6 code. A value of -1 indicates all ICMP/ICMPv6 codes.

securityGroupRuleDescription_fromPort :: Lens' SecurityGroupRuleDescription (Maybe Natural)

The start of the port range for the TCP and UDP protocols, or an ICMP/ICMPv6 type number. A value of -1 indicates all ICMP/ICMPv6 types.

securityGroupRuleDescription_iPV4Range :: Lens' SecurityGroupRuleDescription (Maybe Text)

The IPv4 ranges for the security group rule.

securityGroupRuleDescription_prefixListId :: Lens' SecurityGroupRuleDescription (Maybe Text)

The ID of the prefix list for the security group rule.

securityGroupRuleDescription_protocol :: Lens' SecurityGroupRuleDescription (Maybe Text)

The IP protocol name (tcp, udp, icmp, icmpv6) or number.

securityGroupRuleDescription_iPV6Range :: Lens' SecurityGroupRuleDescription (Maybe Text)

The IPv6 ranges for the security group rule.

securityGroupRuleDescription_toPort :: Lens' SecurityGroupRuleDescription (Maybe Natural)

The end of the port range for the TCP and UDP protocols, or an ICMP/ICMPv6 code. A value of -1 indicates all ICMP/ICMPv6 codes.