Network.AWS.FMS.Types.ViolationDetail

Description

Documentation

data ViolationDetail

Violations for a resource based on the specified Firewall Manager policy and Amazon Web Services account.

See: newViolationDetail smart constructor.

Constructors

ViolationDetail'

Fields

resourceTags :: Maybe [Tag]

The ResourceTag objects associated with the resource.

resourceDescription :: Maybe Text

Brief description for the requested resource.

policyId :: Text

The ID of the Firewall Manager policy that the violation details were requested for.

memberAccount :: Text

The Amazon Web Services account that the violation details were requested for.

resourceId :: Text

The resource ID that the violation details were requested for.

resourceType :: Text

The resource type that the violation details were requested for.

resourceViolations :: [ResourceViolation]

List of violations for the requested resource.

newViolationDetail

Arguments

:: Text
$sel:policyId:ViolationDetail'

-> Text
$sel:memberAccount:ViolationDetail'

-> Text
$sel:resourceId:ViolationDetail'

-> Text
$sel:resourceType:ViolationDetail'

-> ViolationDetail

Create a value of ViolationDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceTags:ViolationDetail', violationDetail_resourceTags - The ResourceTag objects associated with the resource.

$sel:resourceDescription:ViolationDetail', violationDetail_resourceDescription - Brief description for the requested resource.

$sel:policyId:ViolationDetail', violationDetail_policyId - The ID of the Firewall Manager policy that the violation details were requested for.

$sel:memberAccount:ViolationDetail', violationDetail_memberAccount - The Amazon Web Services account that the violation details were requested for.

$sel:resourceId:ViolationDetail', violationDetail_resourceId - The resource ID that the violation details were requested for.

$sel:resourceType:ViolationDetail', violationDetail_resourceType - The resource type that the violation details were requested for.

$sel:resourceViolations:ViolationDetail', violationDetail_resourceViolations - List of violations for the requested resource.

violationDetail_resourceTags :: Lens' ViolationDetail (Maybe [Tag])

The ResourceTag objects associated with the resource.

violationDetail_resourceDescription :: Lens' ViolationDetail (Maybe Text)

Brief description for the requested resource.

violationDetail_policyId :: Lens' ViolationDetail Text

The ID of the Firewall Manager policy that the violation details were requested for.

violationDetail_memberAccount :: Lens' ViolationDetail Text

The Amazon Web Services account that the violation details were requested for.

violationDetail_resourceId :: Lens' ViolationDetail Text

The resource ID that the violation details were requested for.

violationDetail_resourceType :: Lens' ViolationDetail Text

The resource type that the violation details were requested for.

violationDetail_resourceViolations :: Lens' ViolationDetail [ResourceViolation]

List of violations for the requested resource.