Network.AWS.FinSpace

Description

Derived from API version 2021-03-12 of the AWS service descriptions, licensed under Apache 2.0.

The FinSpace management service provides the APIs for managing the FinSpace environments.

Service Configuration

defaultService :: Service

API version 2021-03-12 of the Amazon FinSpace User Environment Management service SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by FinSpace.

ValidationException

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The input fails to satisfy the constraints specified by an AWS service.

AccessDeniedException

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

You do not have sufficient access to perform this action.

InvalidRequestException

_InvalidRequestException :: AsError a => Getting (First ServiceError) a ServiceError

The request is invalid. Something is wrong with the input to the request.

ServiceQuotaExceededException

_ServiceQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

You have exceeded your service quota. To perform the requested action, remove some of the relevant resources, or use Service Quotas to request a service quota increase.

ThrottlingException

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The request was denied due to request throttling.

InternalServerException

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

The request processing has failed because of an unknown error, exception or failure.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

One or more resources can't be found.

LimitExceededException

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

A service limit or quota is exceeded.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

ListEnvironments

data ListEnvironments

See: newListEnvironments smart constructor.

Constructors

ListEnvironments' (Maybe Text) (Maybe Natural)

newListEnvironments :: ListEnvironments

Create a value of ListEnvironments with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListEnvironments', listEnvironments_nextToken - A token generated by FinSpace that specifies where to continue pagination if a previous request was truncated. To get the next set of pages, pass in the nextToken value from the response object of the previous page call.

$sel:maxResults:ListEnvironments', listEnvironments_maxResults - The maximum number of results to return in this request.

data ListEnvironmentsResponse

See: newListEnvironmentsResponse smart constructor.

Constructors

ListEnvironmentsResponse' (Maybe Text) (Maybe [Environment]) Int

newListEnvironmentsResponse

Arguments

:: Int
$sel:httpStatus:ListEnvironmentsResponse'

-> ListEnvironmentsResponse

Create a value of ListEnvironmentsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListEnvironments', listEnvironmentsResponse_nextToken - A token that you can use in a subsequent call to retrieve the next set of results.

$sel:environments:ListEnvironmentsResponse', listEnvironmentsResponse_environments - A list of all of your FinSpace environments.

$sel:httpStatus:ListEnvironmentsResponse', listEnvironmentsResponse_httpStatus - The response's http status code.

UpdateEnvironment

data UpdateEnvironment

See: newUpdateEnvironment smart constructor.

Constructors

UpdateEnvironment' (Maybe FederationParameters) (Maybe FederationMode) (Maybe Text) (Maybe Text) Text

newUpdateEnvironment

Arguments

:: Text
$sel:environmentId:UpdateEnvironment'

-> UpdateEnvironment

Create a value of UpdateEnvironment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:federationParameters:UpdateEnvironment', updateEnvironment_federationParameters - Undocumented member.

$sel:federationMode:UpdateEnvironment', updateEnvironment_federationMode - Authentication mode for the environment.

FEDERATED - Users access FinSpace through Single Sign On (SSO) via your Identity provider.

LOCAL - Users access FinSpace via email and password managed within the FinSpace environment.

$sel:name:UpdateEnvironment', updateEnvironment_name - The name of the environment.

$sel:description:UpdateEnvironment', updateEnvironment_description - The description of the environment.

$sel:environmentId:UpdateEnvironment', updateEnvironment_environmentId - The identifier of the FinSpace environment.

data UpdateEnvironmentResponse

See: newUpdateEnvironmentResponse smart constructor.

Constructors

UpdateEnvironmentResponse' (Maybe Environment) Int

newUpdateEnvironmentResponse

Arguments

:: Int
$sel:httpStatus:UpdateEnvironmentResponse'

-> UpdateEnvironmentResponse

Create a value of UpdateEnvironmentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:environment:UpdateEnvironmentResponse', updateEnvironmentResponse_environment - Returns the FinSpace environment object.

$sel:httpStatus:UpdateEnvironmentResponse', updateEnvironmentResponse_httpStatus - The response's http status code.

DeleteEnvironment

data DeleteEnvironment

See: newDeleteEnvironment smart constructor.

Constructors

DeleteEnvironment' Text

newDeleteEnvironment

Arguments

:: Text
$sel:environmentId:DeleteEnvironment'

-> DeleteEnvironment

Create a value of DeleteEnvironment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:environmentId:DeleteEnvironment', deleteEnvironment_environmentId - The identifier for the FinSpace environment.

data DeleteEnvironmentResponse

See: newDeleteEnvironmentResponse smart constructor.

Constructors

DeleteEnvironmentResponse' Int

newDeleteEnvironmentResponse

Arguments

:: Int
$sel:httpStatus:DeleteEnvironmentResponse'

-> DeleteEnvironmentResponse

Create a value of DeleteEnvironmentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteEnvironmentResponse', deleteEnvironmentResponse_httpStatus - The response's http status code.

ListTagsForResource

data ListTagsForResource

See: newListTagsForResource smart constructor.

Constructors

ListTagsForResource' Text

newListTagsForResource

Arguments

:: Text
$sel:resourceArn:ListTagsForResource'

-> ListTagsForResource

Create a value of ListTagsForResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:ListTagsForResource', listTagsForResource_resourceArn - The Amazon Resource Name of the resource.

data ListTagsForResourceResponse

See: newListTagsForResourceResponse smart constructor.

Constructors

ListTagsForResourceResponse' (Maybe (HashMap Text Text)) Int

newListTagsForResourceResponse

Arguments

:: Int
$sel:httpStatus:ListTagsForResourceResponse'

-> ListTagsForResourceResponse

Create a value of ListTagsForResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:ListTagsForResourceResponse', listTagsForResourceResponse_tags - A list of all tags for a resource.

$sel:httpStatus:ListTagsForResourceResponse', listTagsForResourceResponse_httpStatus - The response's http status code.

TagResource

data TagResource

See: newTagResource smart constructor.

Constructors

TagResource' Text (HashMap Text Text)

newTagResource

Arguments

:: Text
$sel:resourceArn:TagResource'

-> TagResource

Create a value of TagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:TagResource', tagResource_resourceArn - The Amazon Resource Name (ARN) for the resource.

$sel:tags:TagResource', tagResource_tags - One or more tags to be assigned to the resource.

data TagResourceResponse

See: newTagResourceResponse smart constructor.

Constructors

TagResourceResponse' Int

newTagResourceResponse

Arguments

:: Int
$sel:httpStatus:TagResourceResponse'

-> TagResourceResponse

Create a value of TagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:TagResourceResponse', tagResourceResponse_httpStatus - The response's http status code.

GetEnvironment

data GetEnvironment

See: newGetEnvironment smart constructor.

Constructors

GetEnvironment' Text

newGetEnvironment

Arguments

:: Text
$sel:environmentId:GetEnvironment'

-> GetEnvironment

Create a value of GetEnvironment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:environmentId:GetEnvironment', getEnvironment_environmentId - The identifier of the FinSpace environment.

data GetEnvironmentResponse

See: newGetEnvironmentResponse smart constructor.

Constructors

GetEnvironmentResponse' (Maybe Environment) Int

newGetEnvironmentResponse

Arguments

:: Int
$sel:httpStatus:GetEnvironmentResponse'

-> GetEnvironmentResponse

Create a value of GetEnvironmentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:environment:GetEnvironmentResponse', getEnvironmentResponse_environment - The name of the FinSpace environment.

$sel:httpStatus:GetEnvironmentResponse', getEnvironmentResponse_httpStatus - The response's http status code.

UntagResource

data UntagResource

See: newUntagResource smart constructor.

Constructors

UntagResource' Text (NonEmpty Text)

newUntagResource

Arguments

:: Text
$sel:resourceArn:UntagResource'

-> NonEmpty Text
$sel:tagKeys:UntagResource'

-> UntagResource

Create a value of UntagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:UntagResource', untagResource_resourceArn - A FinSpace resource from which you want to remove a tag or tags. The value for this parameter is an Amazon Resource Name (ARN).

$sel:tagKeys:UntagResource', untagResource_tagKeys - The tag keys (names) of one or more tags to be removed.

data UntagResourceResponse

See: newUntagResourceResponse smart constructor.

Constructors

UntagResourceResponse' Int

newUntagResourceResponse

Arguments

:: Int
$sel:httpStatus:UntagResourceResponse'

-> UntagResourceResponse

Create a value of UntagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UntagResourceResponse', untagResourceResponse_httpStatus - The response's http status code.

CreateEnvironment

data CreateEnvironment

See: newCreateEnvironment smart constructor.

Constructors

CreateEnvironment' (Maybe FederationParameters) (Maybe FederationMode) (Maybe Text) (Maybe Text) (Maybe (HashMap Text Text)) Text

newCreateEnvironment

Arguments

:: Text
$sel:name:CreateEnvironment'

-> CreateEnvironment

Create a value of CreateEnvironment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:federationParameters:CreateEnvironment', createEnvironment_federationParameters - Configuration information when authentication mode is FEDERATED.

$sel:federationMode:CreateEnvironment', createEnvironment_federationMode - Authentication mode for the environment.

FEDERATED - Users access FinSpace through Single Sign On (SSO) via your Identity provider.

LOCAL - Users access FinSpace via email and password managed within the FinSpace environment.

$sel:kmsKeyId:CreateEnvironment', createEnvironment_kmsKeyId - The KMS key id to encrypt your data in the FinSpace environment.

$sel:description:CreateEnvironment', createEnvironment_description - The description of the FinSpace environment to be created.

$sel:tags:CreateEnvironment', createEnvironment_tags - Add tags to your FinSpace environment.

$sel:name:CreateEnvironment', createEnvironment_name - The name of the FinSpace environment to be created.

data CreateEnvironmentResponse

See: newCreateEnvironmentResponse smart constructor.

Constructors

CreateEnvironmentResponse' (Maybe Text) (Maybe Text) (Maybe Text) Int

newCreateEnvironmentResponse

Arguments

:: Int
$sel:httpStatus:CreateEnvironmentResponse'

-> CreateEnvironmentResponse

Create a value of CreateEnvironmentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:environmentUrl:CreateEnvironmentResponse', createEnvironmentResponse_environmentUrl - The sign-in url for the web application of the FinSpace environment you created.

$sel:environmentId:CreateEnvironmentResponse', createEnvironmentResponse_environmentId - The unique identifier for FinSpace environment that you created.

$sel:environmentArn:CreateEnvironmentResponse', createEnvironmentResponse_environmentArn - The Amazon Resource Name (ARN) of the FinSpace environment that you created.

$sel:httpStatus:CreateEnvironmentResponse', createEnvironmentResponse_httpStatus - The response's http status code.

Types

EnvironmentStatus

newtype EnvironmentStatus

Constructors

EnvironmentStatus'

Fields

fromEnvironmentStatus :: Text

Bundled Patterns

pattern EnvironmentStatus_CREATED :: EnvironmentStatus

pattern EnvironmentStatus_CREATE_REQUESTED :: EnvironmentStatus

pattern EnvironmentStatus_CREATING :: EnvironmentStatus

pattern EnvironmentStatus_DELETED :: EnvironmentStatus

pattern EnvironmentStatus_DELETE_REQUESTED :: EnvironmentStatus

pattern EnvironmentStatus_DELETING :: EnvironmentStatus

pattern EnvironmentStatus_FAILED_CREATION :: EnvironmentStatus

pattern EnvironmentStatus_FAILED_DELETION :: EnvironmentStatus

pattern EnvironmentStatus_RETRY_DELETION :: EnvironmentStatus

pattern EnvironmentStatus_SUSPENDED :: EnvironmentStatus

FederationMode

newtype FederationMode

Constructors

FederationMode'

Fields

fromFederationMode :: Text

Bundled Patterns

pattern FederationMode_FEDERATED :: FederationMode

pattern FederationMode_LOCAL :: FederationMode

Environment

data Environment

Represents an FinSpace environment.

See: newEnvironment smart constructor.

Constructors

Environment' (Maybe EnvironmentStatus) (Maybe FederationParameters) (Maybe Text) (Maybe Text) (Maybe FederationMode) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text)

newEnvironment :: Environment

Create a value of Environment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Environment', environment_status - The current status of creation of the FinSpace environment.

$sel:federationParameters:Environment', environment_federationParameters - Configuration information when authentication mode is FEDERATED.

$sel:dedicatedServiceAccountId:Environment', environment_dedicatedServiceAccountId - The AWS account ID of the dedicated service account associated with your FinSpace environment.

$sel:environmentUrl:Environment', environment_environmentUrl - The sign-in url for the web application of your FinSpace environment.

$sel:federationMode:Environment', environment_federationMode - The authentication mode for the environment.

$sel:awsAccountId:Environment', environment_awsAccountId - The ID of the AWS account in which the FinSpace environment is created.

$sel:name:Environment', environment_name - The name of the FinSpace environment.

$sel:kmsKeyId:Environment', environment_kmsKeyId - The KMS key id used to encrypt in the FinSpace environment.

$sel:environmentId:Environment', environment_environmentId - The identifier of the FinSpace environment.

$sel:environmentArn:Environment', environment_environmentArn - The Amazon Resource Name (ARN) of your FinSpace environment.

$sel:sageMakerStudioDomainUrl:Environment', environment_sageMakerStudioDomainUrl - The url of the integrated FinSpace notebook environment in your web application.

$sel:description:Environment', environment_description - The description of the FinSpace environment.

FederationParameters

data FederationParameters

Configuration information when authentication mode is FEDERATED.

See: newFederationParameters smart constructor.

Constructors

FederationParameters' (Maybe Text) (Maybe Text) (Maybe Text) (Maybe (HashMap Text Text)) (Maybe Text) (Maybe Text)

newFederationParameters :: FederationParameters

Create a value of FederationParameters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:samlMetadataURL:FederationParameters', federationParameters_samlMetadataURL - Provide the metadata URL from your SAML 2.0 compliant identity provider (IdP).

$sel:applicationCallBackURL:FederationParameters', federationParameters_applicationCallBackURL - The redirect or sign-in URL that should be entered into the SAML 2.0 compliant identity provider configuration (IdP).

$sel:federationURN:FederationParameters', federationParameters_federationURN - The Uniform Resource Name (URN). Also referred as Service Provider URN or Audience URI or Service Provider Entity ID.

$sel:attributeMap:FederationParameters', federationParameters_attributeMap - SAML attribute name and value. The name must always be Email and the value should be set to the attribute definition in which user email is set. For example, name would be Email and value http://schemas.xmlsoap.org/ws/2005/05/identity/claims/emailaddress. Please check your SAML 2.0 compliant identity provider (IdP) documentation for details.

$sel:federationProviderName:FederationParameters', federationParameters_federationProviderName - Name of the identity provider (IdP).

$sel:samlMetadataDocument:FederationParameters', federationParameters_samlMetadataDocument - SAML 2.0 Metadata document from identity provider (IdP).