Network.AWS.FinSpace.Types

Description

Service Configuration

defaultService :: Service

API version 2021-03-12 of the Amazon FinSpace User Environment Management service SDK configuration.

Errors

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The input fails to satisfy the constraints specified by an AWS service.

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

You do not have sufficient access to perform this action.

_InvalidRequestException :: AsError a => Getting (First ServiceError) a ServiceError

The request is invalid. Something is wrong with the input to the request.

_ServiceQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

You have exceeded your service quota. To perform the requested action, remove some of the relevant resources, or use Service Quotas to request a service quota increase.

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The request was denied due to request throttling.

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

The request processing has failed because of an unknown error, exception or failure.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

One or more resources can't be found.

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

A service limit or quota is exceeded.

EnvironmentStatus

newtype EnvironmentStatus

Constructors

EnvironmentStatus'

Fields

fromEnvironmentStatus :: Text

Bundled Patterns

pattern EnvironmentStatus_CREATED :: EnvironmentStatus

pattern EnvironmentStatus_CREATE_REQUESTED :: EnvironmentStatus

pattern EnvironmentStatus_CREATING :: EnvironmentStatus

pattern EnvironmentStatus_DELETED :: EnvironmentStatus

pattern EnvironmentStatus_DELETE_REQUESTED :: EnvironmentStatus

pattern EnvironmentStatus_DELETING :: EnvironmentStatus

pattern EnvironmentStatus_FAILED_CREATION :: EnvironmentStatus

pattern EnvironmentStatus_FAILED_DELETION :: EnvironmentStatus

pattern EnvironmentStatus_RETRY_DELETION :: EnvironmentStatus

pattern EnvironmentStatus_SUSPENDED :: EnvironmentStatus

FederationMode

newtype FederationMode

Constructors

FederationMode'

Fields

fromFederationMode :: Text

Bundled Patterns

pattern FederationMode_FEDERATED :: FederationMode

pattern FederationMode_LOCAL :: FederationMode

Environment

data Environment

Represents an FinSpace environment.

See: newEnvironment smart constructor.

Constructors

Environment'

Fields

status :: Maybe EnvironmentStatus

The current status of creation of the FinSpace environment.

federationParameters :: Maybe FederationParameters

Configuration information when authentication mode is FEDERATED.

dedicatedServiceAccountId :: Maybe Text

The AWS account ID of the dedicated service account associated with your FinSpace environment.

environmentUrl :: Maybe Text

The sign-in url for the web application of your FinSpace environment.

federationMode :: Maybe FederationMode

The authentication mode for the environment.

awsAccountId :: Maybe Text

The ID of the AWS account in which the FinSpace environment is created.

name :: Maybe Text

The name of the FinSpace environment.

kmsKeyId :: Maybe Text

The KMS key id used to encrypt in the FinSpace environment.

environmentId :: Maybe Text

The identifier of the FinSpace environment.

environmentArn :: Maybe Text

The Amazon Resource Name (ARN) of your FinSpace environment.

sageMakerStudioDomainUrl :: Maybe Text

The url of the integrated FinSpace notebook environment in your web application.

description :: Maybe Text

The description of the FinSpace environment.

newEnvironment :: Environment

Create a value of Environment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Environment', environment_status - The current status of creation of the FinSpace environment.

$sel:federationParameters:Environment', environment_federationParameters - Configuration information when authentication mode is FEDERATED.

$sel:dedicatedServiceAccountId:Environment', environment_dedicatedServiceAccountId - The AWS account ID of the dedicated service account associated with your FinSpace environment.

$sel:environmentUrl:Environment', environment_environmentUrl - The sign-in url for the web application of your FinSpace environment.

$sel:federationMode:Environment', environment_federationMode - The authentication mode for the environment.

$sel:awsAccountId:Environment', environment_awsAccountId - The ID of the AWS account in which the FinSpace environment is created.

$sel:name:Environment', environment_name - The name of the FinSpace environment.

$sel:kmsKeyId:Environment', environment_kmsKeyId - The KMS key id used to encrypt in the FinSpace environment.

$sel:environmentId:Environment', environment_environmentId - The identifier of the FinSpace environment.

$sel:environmentArn:Environment', environment_environmentArn - The Amazon Resource Name (ARN) of your FinSpace environment.

$sel:sageMakerStudioDomainUrl:Environment', environment_sageMakerStudioDomainUrl - The url of the integrated FinSpace notebook environment in your web application.

$sel:description:Environment', environment_description - The description of the FinSpace environment.

environment_status :: Lens' Environment (Maybe EnvironmentStatus)

The current status of creation of the FinSpace environment.

environment_federationParameters :: Lens' Environment (Maybe FederationParameters)

Configuration information when authentication mode is FEDERATED.

environment_dedicatedServiceAccountId :: Lens' Environment (Maybe Text)

The AWS account ID of the dedicated service account associated with your FinSpace environment.

environment_environmentUrl :: Lens' Environment (Maybe Text)

The sign-in url for the web application of your FinSpace environment.

environment_federationMode :: Lens' Environment (Maybe FederationMode)

The authentication mode for the environment.

environment_awsAccountId :: Lens' Environment (Maybe Text)

The ID of the AWS account in which the FinSpace environment is created.

environment_name :: Lens' Environment (Maybe Text)

The name of the FinSpace environment.

environment_kmsKeyId :: Lens' Environment (Maybe Text)

The KMS key id used to encrypt in the FinSpace environment.

environment_environmentId :: Lens' Environment (Maybe Text)

The identifier of the FinSpace environment.

environment_environmentArn :: Lens' Environment (Maybe Text)

The Amazon Resource Name (ARN) of your FinSpace environment.

environment_sageMakerStudioDomainUrl :: Lens' Environment (Maybe Text)

The url of the integrated FinSpace notebook environment in your web application.

environment_description :: Lens' Environment (Maybe Text)

The description of the FinSpace environment.

FederationParameters

data FederationParameters

Configuration information when authentication mode is FEDERATED.

See: newFederationParameters smart constructor.

Constructors

FederationParameters'

Fields

samlMetadataURL :: Maybe Text

Provide the metadata URL from your SAML 2.0 compliant identity provider (IdP).

applicationCallBackURL :: Maybe Text

The redirect or sign-in URL that should be entered into the SAML 2.0 compliant identity provider configuration (IdP).

federationURN :: Maybe Text

The Uniform Resource Name (URN). Also referred as Service Provider URN or Audience URI or Service Provider Entity ID.

attributeMap :: Maybe (HashMap Text Text)

SAML attribute name and value. The name must always be Email and the value should be set to the attribute definition in which user email is set. For example, name would be Email and value http://schemas.xmlsoap.org/ws/2005/05/identity/claims/emailaddress. Please check your SAML 2.0 compliant identity provider (IdP) documentation for details.

federationProviderName :: Maybe Text

Name of the identity provider (IdP).

samlMetadataDocument :: Maybe Text

SAML 2.0 Metadata document from identity provider (IdP).

newFederationParameters :: FederationParameters

Create a value of FederationParameters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:samlMetadataURL:FederationParameters', federationParameters_samlMetadataURL - Provide the metadata URL from your SAML 2.0 compliant identity provider (IdP).

$sel:applicationCallBackURL:FederationParameters', federationParameters_applicationCallBackURL - The redirect or sign-in URL that should be entered into the SAML 2.0 compliant identity provider configuration (IdP).

$sel:federationURN:FederationParameters', federationParameters_federationURN - The Uniform Resource Name (URN). Also referred as Service Provider URN or Audience URI or Service Provider Entity ID.

$sel:attributeMap:FederationParameters', federationParameters_attributeMap - SAML attribute name and value. The name must always be Email and the value should be set to the attribute definition in which user email is set. For example, name would be Email and value http://schemas.xmlsoap.org/ws/2005/05/identity/claims/emailaddress. Please check your SAML 2.0 compliant identity provider (IdP) documentation for details.

$sel:federationProviderName:FederationParameters', federationParameters_federationProviderName - Name of the identity provider (IdP).

$sel:samlMetadataDocument:FederationParameters', federationParameters_samlMetadataDocument - SAML 2.0 Metadata document from identity provider (IdP).

federationParameters_samlMetadataURL :: Lens' FederationParameters (Maybe Text)

Provide the metadata URL from your SAML 2.0 compliant identity provider (IdP).

federationParameters_applicationCallBackURL :: Lens' FederationParameters (Maybe Text)

The redirect or sign-in URL that should be entered into the SAML 2.0 compliant identity provider configuration (IdP).

federationParameters_federationURN :: Lens' FederationParameters (Maybe Text)

The Uniform Resource Name (URN). Also referred as Service Provider URN or Audience URI or Service Provider Entity ID.

federationParameters_attributeMap :: Lens' FederationParameters (Maybe (HashMap Text Text))

SAML attribute name and value. The name must always be Email and the value should be set to the attribute definition in which user email is set. For example, name would be Email and value http://schemas.xmlsoap.org/ws/2005/05/identity/claims/emailaddress. Please check your SAML 2.0 compliant identity provider (IdP) documentation for details.

federationParameters_federationProviderName :: Lens' FederationParameters (Maybe Text)

Name of the identity provider (IdP).

federationParameters_samlMetadataDocument :: Lens' FederationParameters (Maybe Text)

SAML 2.0 Metadata document from identity provider (IdP).