Network.AWS.FinSpace.Types.Environment

Description

Documentation

data Environment

Represents an FinSpace environment.

See: newEnvironment smart constructor.

Constructors

Environment'

Fields

status :: Maybe EnvironmentStatus

The current status of creation of the FinSpace environment.

federationParameters :: Maybe FederationParameters

Configuration information when authentication mode is FEDERATED.

dedicatedServiceAccountId :: Maybe Text

The AWS account ID of the dedicated service account associated with your FinSpace environment.

environmentUrl :: Maybe Text

The sign-in url for the web application of your FinSpace environment.

federationMode :: Maybe FederationMode

The authentication mode for the environment.

awsAccountId :: Maybe Text

The ID of the AWS account in which the FinSpace environment is created.

name :: Maybe Text

The name of the FinSpace environment.

kmsKeyId :: Maybe Text

The KMS key id used to encrypt in the FinSpace environment.

environmentId :: Maybe Text

The identifier of the FinSpace environment.

environmentArn :: Maybe Text

The Amazon Resource Name (ARN) of your FinSpace environment.

sageMakerStudioDomainUrl :: Maybe Text

The url of the integrated FinSpace notebook environment in your web application.

description :: Maybe Text

The description of the FinSpace environment.

newEnvironment :: Environment

Create a value of Environment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Environment', environment_status - The current status of creation of the FinSpace environment.

$sel:federationParameters:Environment', environment_federationParameters - Configuration information when authentication mode is FEDERATED.

$sel:dedicatedServiceAccountId:Environment', environment_dedicatedServiceAccountId - The AWS account ID of the dedicated service account associated with your FinSpace environment.

$sel:environmentUrl:Environment', environment_environmentUrl - The sign-in url for the web application of your FinSpace environment.

$sel:federationMode:Environment', environment_federationMode - The authentication mode for the environment.

$sel:awsAccountId:Environment', environment_awsAccountId - The ID of the AWS account in which the FinSpace environment is created.

$sel:name:Environment', environment_name - The name of the FinSpace environment.

$sel:kmsKeyId:Environment', environment_kmsKeyId - The KMS key id used to encrypt in the FinSpace environment.

$sel:environmentId:Environment', environment_environmentId - The identifier of the FinSpace environment.

$sel:environmentArn:Environment', environment_environmentArn - The Amazon Resource Name (ARN) of your FinSpace environment.

$sel:sageMakerStudioDomainUrl:Environment', environment_sageMakerStudioDomainUrl - The url of the integrated FinSpace notebook environment in your web application.

$sel:description:Environment', environment_description - The description of the FinSpace environment.

environment_status :: Lens' Environment (Maybe EnvironmentStatus)

The current status of creation of the FinSpace environment.

environment_federationParameters :: Lens' Environment (Maybe FederationParameters)

Configuration information when authentication mode is FEDERATED.

environment_dedicatedServiceAccountId :: Lens' Environment (Maybe Text)

The AWS account ID of the dedicated service account associated with your FinSpace environment.

environment_environmentUrl :: Lens' Environment (Maybe Text)

The sign-in url for the web application of your FinSpace environment.

environment_federationMode :: Lens' Environment (Maybe FederationMode)

The authentication mode for the environment.

environment_awsAccountId :: Lens' Environment (Maybe Text)

The ID of the AWS account in which the FinSpace environment is created.

environment_name :: Lens' Environment (Maybe Text)

The name of the FinSpace environment.

environment_kmsKeyId :: Lens' Environment (Maybe Text)

The KMS key id used to encrypt in the FinSpace environment.

environment_environmentId :: Lens' Environment (Maybe Text)

The identifier of the FinSpace environment.

environment_environmentArn :: Lens' Environment (Maybe Text)

The Amazon Resource Name (ARN) of your FinSpace environment.

environment_sageMakerStudioDomainUrl :: Lens' Environment (Maybe Text)

The url of the integrated FinSpace notebook environment in your web application.

environment_description :: Lens' Environment (Maybe Text)

The description of the FinSpace environment.