Network.AWS.FinSpace.Types.FederationParameters

Description

Documentation

data FederationParameters

Configuration information when authentication mode is FEDERATED.

See: newFederationParameters smart constructor.

Constructors

FederationParameters'

Fields

samlMetadataURL :: Maybe Text

Provide the metadata URL from your SAML 2.0 compliant identity provider (IdP).

applicationCallBackURL :: Maybe Text

The redirect or sign-in URL that should be entered into the SAML 2.0 compliant identity provider configuration (IdP).

federationURN :: Maybe Text

The Uniform Resource Name (URN). Also referred as Service Provider URN or Audience URI or Service Provider Entity ID.

attributeMap :: Maybe (HashMap Text Text)

SAML attribute name and value. The name must always be Email and the value should be set to the attribute definition in which user email is set. For example, name would be Email and value http://schemas.xmlsoap.org/ws/2005/05/identity/claims/emailaddress. Please check your SAML 2.0 compliant identity provider (IdP) documentation for details.

federationProviderName :: Maybe Text

Name of the identity provider (IdP).

samlMetadataDocument :: Maybe Text

SAML 2.0 Metadata document from identity provider (IdP).

newFederationParameters :: FederationParameters

Create a value of FederationParameters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:samlMetadataURL:FederationParameters', federationParameters_samlMetadataURL - Provide the metadata URL from your SAML 2.0 compliant identity provider (IdP).

$sel:applicationCallBackURL:FederationParameters', federationParameters_applicationCallBackURL - The redirect or sign-in URL that should be entered into the SAML 2.0 compliant identity provider configuration (IdP).

$sel:federationURN:FederationParameters', federationParameters_federationURN - The Uniform Resource Name (URN). Also referred as Service Provider URN or Audience URI or Service Provider Entity ID.

$sel:attributeMap:FederationParameters', federationParameters_attributeMap - SAML attribute name and value. The name must always be Email and the value should be set to the attribute definition in which user email is set. For example, name would be Email and value http://schemas.xmlsoap.org/ws/2005/05/identity/claims/emailaddress. Please check your SAML 2.0 compliant identity provider (IdP) documentation for details.

$sel:federationProviderName:FederationParameters', federationParameters_federationProviderName - Name of the identity provider (IdP).

$sel:samlMetadataDocument:FederationParameters', federationParameters_samlMetadataDocument - SAML 2.0 Metadata document from identity provider (IdP).

federationParameters_samlMetadataURL :: Lens' FederationParameters (Maybe Text)

Provide the metadata URL from your SAML 2.0 compliant identity provider (IdP).

federationParameters_applicationCallBackURL :: Lens' FederationParameters (Maybe Text)

The redirect or sign-in URL that should be entered into the SAML 2.0 compliant identity provider configuration (IdP).

federationParameters_federationURN :: Lens' FederationParameters (Maybe Text)

The Uniform Resource Name (URN). Also referred as Service Provider URN or Audience URI or Service Provider Entity ID.

federationParameters_attributeMap :: Lens' FederationParameters (Maybe (HashMap Text Text))

SAML attribute name and value. The name must always be Email and the value should be set to the attribute definition in which user email is set. For example, name would be Email and value http://schemas.xmlsoap.org/ws/2005/05/identity/claims/emailaddress. Please check your SAML 2.0 compliant identity provider (IdP) documentation for details.

federationParameters_federationProviderName :: Lens' FederationParameters (Maybe Text)

Name of the identity provider (IdP).

federationParameters_samlMetadataDocument :: Lens' FederationParameters (Maybe Text)

SAML 2.0 Metadata document from identity provider (IdP).