Network.AWS.FinSpace.UpdateEnvironment

Description

Update your FinSpace environment.

Creating a Request

data UpdateEnvironment

See: newUpdateEnvironment smart constructor.

Constructors

UpdateEnvironment'

Fields

federationParameters :: Maybe FederationParameters

federationMode :: Maybe FederationMode

Authentication mode for the environment.

FEDERATED - Users access FinSpace through Single Sign On (SSO) via your Identity provider.

LOCAL - Users access FinSpace via email and password managed within the FinSpace environment.

name :: Maybe Text

The name of the environment.

description :: Maybe Text

The description of the environment.

environmentId :: Text

The identifier of the FinSpace environment.

newUpdateEnvironment

Arguments

:: Text
$sel:environmentId:UpdateEnvironment'

-> UpdateEnvironment

Create a value of UpdateEnvironment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:federationParameters:UpdateEnvironment', updateEnvironment_federationParameters - Undocumented member.

$sel:federationMode:UpdateEnvironment', updateEnvironment_federationMode - Authentication mode for the environment.

FEDERATED - Users access FinSpace through Single Sign On (SSO) via your Identity provider.

LOCAL - Users access FinSpace via email and password managed within the FinSpace environment.

$sel:name:UpdateEnvironment', updateEnvironment_name - The name of the environment.

$sel:description:UpdateEnvironment', updateEnvironment_description - The description of the environment.

$sel:environmentId:UpdateEnvironment', updateEnvironment_environmentId - The identifier of the FinSpace environment.

Request Lenses

updateEnvironment_federationParameters :: Lens' UpdateEnvironment (Maybe FederationParameters)

Undocumented member.

updateEnvironment_federationMode :: Lens' UpdateEnvironment (Maybe FederationMode)

Authentication mode for the environment.

FEDERATED - Users access FinSpace through Single Sign On (SSO) via your Identity provider.

LOCAL - Users access FinSpace via email and password managed within the FinSpace environment.

updateEnvironment_name :: Lens' UpdateEnvironment (Maybe Text)

The name of the environment.

updateEnvironment_description :: Lens' UpdateEnvironment (Maybe Text)

The description of the environment.

updateEnvironment_environmentId :: Lens' UpdateEnvironment Text

The identifier of the FinSpace environment.

Destructuring the Response

data UpdateEnvironmentResponse

See: newUpdateEnvironmentResponse smart constructor.

Constructors

UpdateEnvironmentResponse'

Fields

environment :: Maybe Environment

Returns the FinSpace environment object.

httpStatus :: Int

The response's http status code.

newUpdateEnvironmentResponse

Arguments

:: Int
$sel:httpStatus:UpdateEnvironmentResponse'

-> UpdateEnvironmentResponse

Create a value of UpdateEnvironmentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:environment:UpdateEnvironmentResponse', updateEnvironmentResponse_environment - Returns the FinSpace environment object.

$sel:httpStatus:UpdateEnvironmentResponse', updateEnvironmentResponse_httpStatus - The response's http status code.

Response Lenses

updateEnvironmentResponse_environment :: Lens' UpdateEnvironmentResponse (Maybe Environment)

Returns the FinSpace environment object.

updateEnvironmentResponse_httpStatus :: Lens' UpdateEnvironmentResponse Int

The response's http status code.