Network.AWS.FinSpaceData.CreateChangeset

Description

Creates a new changeset in a FinSpace dataset.

Creating a Request

data CreateChangeset

See: newCreateChangeset smart constructor.

Constructors

CreateChangeset'

Fields

formatParams :: Maybe (HashMap Text Text)

Options that define the structure of the source file(s).

formatType :: Maybe FormatType

Format type of the input files being loaded into the changeset.

tags :: Maybe (HashMap Text Text)

Metadata tags to apply to this changeset.

datasetId :: Text

The unique identifier for the FinSpace dataset in which the changeset will be created.

changeType :: ChangeType

Option to indicate how a changeset will be applied to a dataset.

REPLACE - Changeset will be considered as a replacement to all prior loaded changesets.

APPEND - Changeset will be considered as an addition to the end of all prior loaded changesets.

sourceType :: SourceType

Type of the data source from which the files to create the changeset will be sourced.

S3 - Amazon S3.

sourceParams :: HashMap Text Text

Source path from which the files to create the changeset will be sourced.

newCreateChangeset

Arguments

:: Text
$sel:datasetId:CreateChangeset'

-> ChangeType
$sel:changeType:CreateChangeset'

-> SourceType
$sel:sourceType:CreateChangeset'

-> CreateChangeset

Create a value of CreateChangeset with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:formatParams:CreateChangeset', createChangeset_formatParams - Options that define the structure of the source file(s).

$sel:formatType:CreateChangeset', createChangeset_formatType - Format type of the input files being loaded into the changeset.

$sel:tags:CreateChangeset', createChangeset_tags - Metadata tags to apply to this changeset.

$sel:datasetId:CreateChangeset', createChangeset_datasetId - The unique identifier for the FinSpace dataset in which the changeset will be created.

$sel:changeType:CreateChangeset', createChangeset_changeType - Option to indicate how a changeset will be applied to a dataset.

REPLACE - Changeset will be considered as a replacement to all prior loaded changesets.

APPEND - Changeset will be considered as an addition to the end of all prior loaded changesets.

$sel:sourceType:CreateChangeset', createChangeset_sourceType - Type of the data source from which the files to create the changeset will be sourced.

S3 - Amazon S3.

$sel:sourceParams:CreateChangeset', createChangeset_sourceParams - Source path from which the files to create the changeset will be sourced.

Request Lenses

createChangeset_formatParams :: Lens' CreateChangeset (Maybe (HashMap Text Text))

Options that define the structure of the source file(s).

createChangeset_formatType :: Lens' CreateChangeset (Maybe FormatType)

Format type of the input files being loaded into the changeset.

createChangeset_tags :: Lens' CreateChangeset (Maybe (HashMap Text Text))

Metadata tags to apply to this changeset.

createChangeset_datasetId :: Lens' CreateChangeset Text

The unique identifier for the FinSpace dataset in which the changeset will be created.

createChangeset_changeType :: Lens' CreateChangeset ChangeType

Option to indicate how a changeset will be applied to a dataset.

REPLACE - Changeset will be considered as a replacement to all prior loaded changesets.

APPEND - Changeset will be considered as an addition to the end of all prior loaded changesets.

createChangeset_sourceType :: Lens' CreateChangeset SourceType

Type of the data source from which the files to create the changeset will be sourced.

S3 - Amazon S3.

createChangeset_sourceParams :: Lens' CreateChangeset (HashMap Text Text)

Source path from which the files to create the changeset will be sourced.

Destructuring the Response

data CreateChangesetResponse

See: newCreateChangesetResponse smart constructor.

Constructors

CreateChangesetResponse'

Fields

changeset :: Maybe ChangesetInfo

Returns the changeset details.

httpStatus :: Int

The response's http status code.

newCreateChangesetResponse

Arguments

:: Int
$sel:httpStatus:CreateChangesetResponse'

-> CreateChangesetResponse

Create a value of CreateChangesetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:changeset:CreateChangesetResponse', createChangesetResponse_changeset - Returns the changeset details.

$sel:httpStatus:CreateChangesetResponse', createChangesetResponse_httpStatus - The response's http status code.

Response Lenses

createChangesetResponse_changeset :: Lens' CreateChangesetResponse (Maybe ChangesetInfo)

Returns the changeset details.

createChangesetResponse_httpStatus :: Lens' CreateChangesetResponse Int

The response's http status code.