Network.AWS.FinSpaceData.GetProgrammaticAccessCredentials

Description

Request programmatic credentials to use with Habanero SDK.

Creating a Request

data GetProgrammaticAccessCredentials

See: newGetProgrammaticAccessCredentials smart constructor.

Constructors

GetProgrammaticAccessCredentials'

Fields

durationInMinutes :: Maybe Natural

The time duration in which the credentials remain valid.

environmentId :: Text

The habanero environment identifier.

newGetProgrammaticAccessCredentials

Arguments

:: Text
$sel:environmentId:GetProgrammaticAccessCredentials'

-> GetProgrammaticAccessCredentials

Create a value of GetProgrammaticAccessCredentials with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:durationInMinutes:GetProgrammaticAccessCredentials', getProgrammaticAccessCredentials_durationInMinutes - The time duration in which the credentials remain valid.

$sel:environmentId:GetProgrammaticAccessCredentials', getProgrammaticAccessCredentials_environmentId - The habanero environment identifier.

Request Lenses

getProgrammaticAccessCredentials_durationInMinutes :: Lens' GetProgrammaticAccessCredentials (Maybe Natural)

The time duration in which the credentials remain valid.

getProgrammaticAccessCredentials_environmentId :: Lens' GetProgrammaticAccessCredentials Text

The habanero environment identifier.

Destructuring the Response

data GetProgrammaticAccessCredentialsResponse

See: newGetProgrammaticAccessCredentialsResponse smart constructor.

Constructors

GetProgrammaticAccessCredentialsResponse'

Fields

credentials :: Maybe Credentials

Returns the programmatic credentials.

durationInMinutes :: Maybe Natural

Returns the duration in which the credentials will remain valid.

httpStatus :: Int

The response's http status code.

newGetProgrammaticAccessCredentialsResponse

Arguments

:: Int
$sel:httpStatus:GetProgrammaticAccessCredentialsResponse'

-> GetProgrammaticAccessCredentialsResponse

Create a value of GetProgrammaticAccessCredentialsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:credentials:GetProgrammaticAccessCredentialsResponse', getProgrammaticAccessCredentialsResponse_credentials - Returns the programmatic credentials.

$sel:durationInMinutes:GetProgrammaticAccessCredentials', getProgrammaticAccessCredentialsResponse_durationInMinutes - Returns the duration in which the credentials will remain valid.

$sel:httpStatus:GetProgrammaticAccessCredentialsResponse', getProgrammaticAccessCredentialsResponse_httpStatus - The response's http status code.

Response Lenses

getProgrammaticAccessCredentialsResponse_credentials :: Lens' GetProgrammaticAccessCredentialsResponse (Maybe Credentials)

Returns the programmatic credentials.

getProgrammaticAccessCredentialsResponse_durationInMinutes :: Lens' GetProgrammaticAccessCredentialsResponse (Maybe Natural)

Returns the duration in which the credentials will remain valid.

getProgrammaticAccessCredentialsResponse_httpStatus :: Lens' GetProgrammaticAccessCredentialsResponse Int

The response's http status code.