Network.AWS.Firehose.CreateDeliveryStream

Description

Creates a Kinesis Data Firehose delivery stream.

By default, you can create up to 50 delivery streams per AWS Region.

This is an asynchronous operation that immediately returns. The initial status of the delivery stream is CREATING. After the delivery stream is created, its status is ACTIVE and it now accepts data. If the delivery stream creation fails, the status transitions to CREATING_FAILED. Attempts to send data to a delivery stream that is not in the ACTIVE state cause an exception. To check the state of a delivery stream, use DescribeDeliveryStream.

If the status of a delivery stream is CREATING_FAILED, this status doesn't change, and you can't invoke CreateDeliveryStream again on it. However, you can invoke the DeleteDeliveryStream operation to delete it.

A Kinesis Data Firehose delivery stream can be configured to receive records directly from providers using PutRecord or PutRecordBatch, or it can be configured to use an existing Kinesis stream as its source. To specify a Kinesis data stream as input, set the DeliveryStreamType parameter to KinesisStreamAsSource, and provide the Kinesis stream Amazon Resource Name (ARN) and role ARN in the KinesisStreamSourceConfiguration parameter.

To create a delivery stream with server-side encryption (SSE) enabled, include DeliveryStreamEncryptionConfigurationInput in your request. This is optional. You can also invoke StartDeliveryStreamEncryption to turn on SSE for an existing delivery stream that doesn't have SSE enabled.

A delivery stream is configured with a single destination: Amazon S3, Amazon ES, Amazon Redshift, or Splunk. You must specify only one of the following destination configuration parameters: ExtendedS3DestinationConfiguration, S3DestinationConfiguration, ElasticsearchDestinationConfiguration, RedshiftDestinationConfiguration, or SplunkDestinationConfiguration.

When you specify S3DestinationConfiguration, you can also provide the following optional values: BufferingHints, EncryptionConfiguration, and CompressionFormat. By default, if no BufferingHints value is provided, Kinesis Data Firehose buffers data up to 5 MB or for 5 minutes, whichever condition is satisfied first. BufferingHints is a hint, so there are some cases where the service cannot adhere to these conditions strictly. For example, record boundaries might be such that the size is a little over or under the configured buffering size. By default, no encryption is performed. We strongly recommend that you enable encryption to ensure secure data storage in Amazon S3.

A few notes about Amazon Redshift as a destination:

An Amazon Redshift destination requires an S3 bucket as intermediate location. Kinesis Data Firehose first delivers data to Amazon S3 and then uses COPY syntax to load data into an Amazon Redshift table. This is specified in the RedshiftDestinationConfiguration.S3Configuration parameter.

The compression formats SNAPPY or ZIP cannot be specified in RedshiftDestinationConfiguration.S3Configuration because the Amazon Redshift COPY operation that reads from the S3 bucket doesn't support these compression formats.

We strongly recommend that you use the user name and password you provide exclusively with Kinesis Data Firehose, and that the permissions for the account are restricted for Amazon Redshift INSERT permissions.

Kinesis Data Firehose assumes the IAM role that is configured as part of the destination. The role should allow the Kinesis Data Firehose principal to assume the role, and the role should have permissions that allow the service to deliver the data. For more information, see Grant Kinesis Data Firehose Access to an Amazon S3 Destination in the Amazon Kinesis Data Firehose Developer Guide.

Creating a Request

data CreateDeliveryStream

See: newCreateDeliveryStream smart constructor.

Constructors

CreateDeliveryStream'

Fields

s3DestinationConfiguration :: Maybe S3DestinationConfiguration

Deprecated

The destination in Amazon S3. You can specify only one destination.

redshiftDestinationConfiguration :: Maybe RedshiftDestinationConfiguration

The destination in Amazon Redshift. You can specify only one destination.

elasticsearchDestinationConfiguration :: Maybe ElasticsearchDestinationConfiguration

The destination in Amazon ES. You can specify only one destination.

extendedS3DestinationConfiguration :: Maybe ExtendedS3DestinationConfiguration

The destination in Amazon S3. You can specify only one destination.

kinesisStreamSourceConfiguration :: Maybe KinesisStreamSourceConfiguration

When a Kinesis data stream is used as the source for the delivery stream, a KinesisStreamSourceConfiguration containing the Kinesis data stream Amazon Resource Name (ARN) and the role ARN for the source stream.

httpEndpointDestinationConfiguration :: Maybe HttpEndpointDestinationConfiguration

Enables configuring Kinesis Firehose to deliver data to any HTTP endpoint destination. You can specify only one destination.

amazonopensearchserviceDestinationConfiguration :: Maybe AmazonopensearchserviceDestinationConfiguration

deliveryStreamType :: Maybe DeliveryStreamType

The delivery stream type. This parameter can be one of the following values:

DirectPut: Provider applications access the delivery stream directly.

KinesisStreamAsSource: The delivery stream uses a Kinesis data stream as a source.

splunkDestinationConfiguration :: Maybe SplunkDestinationConfiguration

The destination in Splunk. You can specify only one destination.

tags :: Maybe (NonEmpty Tag)

A set of tags to assign to the delivery stream. A tag is a key-value pair that you can define and assign to AWS resources. Tags are metadata. For example, you can add friendly names and descriptions or other types of information that can help you distinguish the delivery stream. For more information about tags, see Using Cost Allocation Tags in the AWS Billing and Cost Management User Guide.

You can specify up to 50 tags when creating a delivery stream.

deliveryStreamEncryptionConfigurationInput :: Maybe DeliveryStreamEncryptionConfigurationInput

Used to specify the type and Amazon Resource Name (ARN) of the KMS key needed for Server-Side Encryption (SSE).

deliveryStreamName :: Text

The name of the delivery stream. This name must be unique per AWS account in the same AWS Region. If the delivery streams are in different accounts or different Regions, you can have multiple delivery streams with the same name.

newCreateDeliveryStream

Arguments

:: Text
$sel:deliveryStreamName:CreateDeliveryStream'

-> CreateDeliveryStream

Create a value of CreateDeliveryStream with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3DestinationConfiguration:CreateDeliveryStream', createDeliveryStream_s3DestinationConfiguration - [Deprecated] The destination in Amazon S3. You can specify only one destination.

$sel:redshiftDestinationConfiguration:CreateDeliveryStream', createDeliveryStream_redshiftDestinationConfiguration - The destination in Amazon Redshift. You can specify only one destination.

$sel:elasticsearchDestinationConfiguration:CreateDeliveryStream', createDeliveryStream_elasticsearchDestinationConfiguration - The destination in Amazon ES. You can specify only one destination.

$sel:extendedS3DestinationConfiguration:CreateDeliveryStream', createDeliveryStream_extendedS3DestinationConfiguration - The destination in Amazon S3. You can specify only one destination.

$sel:kinesisStreamSourceConfiguration:CreateDeliveryStream', createDeliveryStream_kinesisStreamSourceConfiguration - When a Kinesis data stream is used as the source for the delivery stream, a KinesisStreamSourceConfiguration containing the Kinesis data stream Amazon Resource Name (ARN) and the role ARN for the source stream.

$sel:httpEndpointDestinationConfiguration:CreateDeliveryStream', createDeliveryStream_httpEndpointDestinationConfiguration - Enables configuring Kinesis Firehose to deliver data to any HTTP endpoint destination. You can specify only one destination.

$sel:amazonopensearchserviceDestinationConfiguration:CreateDeliveryStream', createDeliveryStream_amazonopensearchserviceDestinationConfiguration - Undocumented member.

$sel:deliveryStreamType:CreateDeliveryStream', createDeliveryStream_deliveryStreamType - The delivery stream type. This parameter can be one of the following values:

DirectPut: Provider applications access the delivery stream directly.

KinesisStreamAsSource: The delivery stream uses a Kinesis data stream as a source.

$sel:splunkDestinationConfiguration:CreateDeliveryStream', createDeliveryStream_splunkDestinationConfiguration - The destination in Splunk. You can specify only one destination.

$sel:tags:CreateDeliveryStream', createDeliveryStream_tags - A set of tags to assign to the delivery stream. A tag is a key-value pair that you can define and assign to AWS resources. Tags are metadata. For example, you can add friendly names and descriptions or other types of information that can help you distinguish the delivery stream. For more information about tags, see Using Cost Allocation Tags in the AWS Billing and Cost Management User Guide.

You can specify up to 50 tags when creating a delivery stream.

$sel:deliveryStreamEncryptionConfigurationInput:CreateDeliveryStream', createDeliveryStream_deliveryStreamEncryptionConfigurationInput - Used to specify the type and Amazon Resource Name (ARN) of the KMS key needed for Server-Side Encryption (SSE).

$sel:deliveryStreamName:CreateDeliveryStream', createDeliveryStream_deliveryStreamName - The name of the delivery stream. This name must be unique per AWS account in the same AWS Region. If the delivery streams are in different accounts or different Regions, you can have multiple delivery streams with the same name.

Request Lenses

createDeliveryStream_s3DestinationConfiguration :: Lens' CreateDeliveryStream (Maybe S3DestinationConfiguration)

Deprecated

The destination in Amazon S3. You can specify only one destination.

createDeliveryStream_redshiftDestinationConfiguration :: Lens' CreateDeliveryStream (Maybe RedshiftDestinationConfiguration)

The destination in Amazon Redshift. You can specify only one destination.

createDeliveryStream_elasticsearchDestinationConfiguration :: Lens' CreateDeliveryStream (Maybe ElasticsearchDestinationConfiguration)

The destination in Amazon ES. You can specify only one destination.

createDeliveryStream_extendedS3DestinationConfiguration :: Lens' CreateDeliveryStream (Maybe ExtendedS3DestinationConfiguration)

The destination in Amazon S3. You can specify only one destination.

createDeliveryStream_kinesisStreamSourceConfiguration :: Lens' CreateDeliveryStream (Maybe KinesisStreamSourceConfiguration)

When a Kinesis data stream is used as the source for the delivery stream, a KinesisStreamSourceConfiguration containing the Kinesis data stream Amazon Resource Name (ARN) and the role ARN for the source stream.

createDeliveryStream_httpEndpointDestinationConfiguration :: Lens' CreateDeliveryStream (Maybe HttpEndpointDestinationConfiguration)

Enables configuring Kinesis Firehose to deliver data to any HTTP endpoint destination. You can specify only one destination.

createDeliveryStream_amazonopensearchserviceDestinationConfiguration :: Lens' CreateDeliveryStream (Maybe AmazonopensearchserviceDestinationConfiguration)

Undocumented member.

createDeliveryStream_deliveryStreamType :: Lens' CreateDeliveryStream (Maybe DeliveryStreamType)

The delivery stream type. This parameter can be one of the following values:

DirectPut: Provider applications access the delivery stream directly.

KinesisStreamAsSource: The delivery stream uses a Kinesis data stream as a source.

createDeliveryStream_splunkDestinationConfiguration :: Lens' CreateDeliveryStream (Maybe SplunkDestinationConfiguration)

The destination in Splunk. You can specify only one destination.

createDeliveryStream_tags :: Lens' CreateDeliveryStream (Maybe (NonEmpty Tag))

A set of tags to assign to the delivery stream. A tag is a key-value pair that you can define and assign to AWS resources. Tags are metadata. For example, you can add friendly names and descriptions or other types of information that can help you distinguish the delivery stream. For more information about tags, see Using Cost Allocation Tags in the AWS Billing and Cost Management User Guide.

You can specify up to 50 tags when creating a delivery stream.

createDeliveryStream_deliveryStreamEncryptionConfigurationInput :: Lens' CreateDeliveryStream (Maybe DeliveryStreamEncryptionConfigurationInput)

Used to specify the type and Amazon Resource Name (ARN) of the KMS key needed for Server-Side Encryption (SSE).

createDeliveryStream_deliveryStreamName :: Lens' CreateDeliveryStream Text

The name of the delivery stream. This name must be unique per AWS account in the same AWS Region. If the delivery streams are in different accounts or different Regions, you can have multiple delivery streams with the same name.

Destructuring the Response

data CreateDeliveryStreamResponse

See: newCreateDeliveryStreamResponse smart constructor.

Constructors

CreateDeliveryStreamResponse'

Fields

deliveryStreamARN :: Maybe Text

The ARN of the delivery stream.

httpStatus :: Int

The response's http status code.

newCreateDeliveryStreamResponse

Arguments

:: Int
$sel:httpStatus:CreateDeliveryStreamResponse'

-> CreateDeliveryStreamResponse

Create a value of CreateDeliveryStreamResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deliveryStreamARN:CreateDeliveryStreamResponse', createDeliveryStreamResponse_deliveryStreamARN - The ARN of the delivery stream.

$sel:httpStatus:CreateDeliveryStreamResponse', createDeliveryStreamResponse_httpStatus - The response's http status code.

Response Lenses

createDeliveryStreamResponse_deliveryStreamARN :: Lens' CreateDeliveryStreamResponse (Maybe Text)

The ARN of the delivery stream.

createDeliveryStreamResponse_httpStatus :: Lens' CreateDeliveryStreamResponse Int

The response's http status code.