Network.AWS.Firehose.ListTagsForDeliveryStream

Description

Lists the tags for the specified delivery stream. This operation has a limit of five transactions per second per account.

Creating a Request

data ListTagsForDeliveryStream

See: newListTagsForDeliveryStream smart constructor.

Constructors

ListTagsForDeliveryStream'

Fields

exclusiveStartTagKey :: Maybe Text

The key to use as the starting point for the list of tags. If you set this parameter, ListTagsForDeliveryStream gets all tags that occur after ExclusiveStartTagKey.

limit :: Maybe Natural

The number of tags to return. If this number is less than the total number of tags associated with the delivery stream, HasMoreTags is set to true in the response. To list additional tags, set ExclusiveStartTagKey to the last key in the response.

deliveryStreamName :: Text

The name of the delivery stream whose tags you want to list.

newListTagsForDeliveryStream

Arguments

:: Text
$sel:deliveryStreamName:ListTagsForDeliveryStream'

-> ListTagsForDeliveryStream

Create a value of ListTagsForDeliveryStream with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:exclusiveStartTagKey:ListTagsForDeliveryStream', listTagsForDeliveryStream_exclusiveStartTagKey - The key to use as the starting point for the list of tags. If you set this parameter, ListTagsForDeliveryStream gets all tags that occur after ExclusiveStartTagKey.

$sel:limit:ListTagsForDeliveryStream', listTagsForDeliveryStream_limit - The number of tags to return. If this number is less than the total number of tags associated with the delivery stream, HasMoreTags is set to true in the response. To list additional tags, set ExclusiveStartTagKey to the last key in the response.

$sel:deliveryStreamName:ListTagsForDeliveryStream', listTagsForDeliveryStream_deliveryStreamName - The name of the delivery stream whose tags you want to list.

Request Lenses

listTagsForDeliveryStream_exclusiveStartTagKey :: Lens' ListTagsForDeliveryStream (Maybe Text)

The key to use as the starting point for the list of tags. If you set this parameter, ListTagsForDeliveryStream gets all tags that occur after ExclusiveStartTagKey.

listTagsForDeliveryStream_limit :: Lens' ListTagsForDeliveryStream (Maybe Natural)

The number of tags to return. If this number is less than the total number of tags associated with the delivery stream, HasMoreTags is set to true in the response. To list additional tags, set ExclusiveStartTagKey to the last key in the response.

listTagsForDeliveryStream_deliveryStreamName :: Lens' ListTagsForDeliveryStream Text

The name of the delivery stream whose tags you want to list.

Destructuring the Response

data ListTagsForDeliveryStreamResponse

See: newListTagsForDeliveryStreamResponse smart constructor.

Constructors

ListTagsForDeliveryStreamResponse'

Fields

httpStatus :: Int

The response's http status code.

tags :: [Tag]

A list of tags associated with DeliveryStreamName, starting with the first tag after ExclusiveStartTagKey and up to the specified Limit.

hasMoreTags :: Bool

If this is true in the response, more tags are available. To list the remaining tags, set ExclusiveStartTagKey to the key of the last tag returned and call ListTagsForDeliveryStream again.

newListTagsForDeliveryStreamResponse

Arguments

:: Int
$sel:httpStatus:ListTagsForDeliveryStreamResponse'

-> Bool
$sel:hasMoreTags:ListTagsForDeliveryStreamResponse'

-> ListTagsForDeliveryStreamResponse

Create a value of ListTagsForDeliveryStreamResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:ListTagsForDeliveryStreamResponse', listTagsForDeliveryStreamResponse_httpStatus - The response's http status code.

$sel:tags:ListTagsForDeliveryStreamResponse', listTagsForDeliveryStreamResponse_tags - A list of tags associated with DeliveryStreamName, starting with the first tag after ExclusiveStartTagKey and up to the specified Limit.

$sel:hasMoreTags:ListTagsForDeliveryStreamResponse', listTagsForDeliveryStreamResponse_hasMoreTags - If this is true in the response, more tags are available. To list the remaining tags, set ExclusiveStartTagKey to the key of the last tag returned and call ListTagsForDeliveryStream again.

Response Lenses

listTagsForDeliveryStreamResponse_httpStatus :: Lens' ListTagsForDeliveryStreamResponse Int

The response's http status code.

listTagsForDeliveryStreamResponse_tags :: Lens' ListTagsForDeliveryStreamResponse [Tag]

A list of tags associated with DeliveryStreamName, starting with the first tag after ExclusiveStartTagKey and up to the specified Limit.

listTagsForDeliveryStreamResponse_hasMoreTags :: Lens' ListTagsForDeliveryStreamResponse Bool

If this is true in the response, more tags are available. To list the remaining tags, set ExclusiveStartTagKey to the key of the last tag returned and call ListTagsForDeliveryStream again.