Network.AWS.Firehose.PutRecord

Description

Writes a single data record into an Amazon Kinesis Data Firehose delivery stream. To write multiple data records into a delivery stream, use PutRecordBatch. Applications using these operations are referred to as producers.

By default, each delivery stream can take in up to 2,000 transactions per second, 5,000 records per second, or 5 MB per second. If you use PutRecord and PutRecordBatch, the limits are an aggregate across these two operations for each delivery stream. For more information about limits and how to request an increase, see Amazon Kinesis Data Firehose Limits.

You must specify the name of the delivery stream and the data record when using PutRecord. The data record consists of a data blob that can be up to 1,000 KiB in size, and any kind of data. For example, it can be a segment from a log file, geographic location data, website clickstream data, and so on.

Kinesis Data Firehose buffers records before delivering them to the destination. To disambiguate the data blobs at the destination, a common solution is to use delimiters in the data, such as a newline (\n) or some other character unique within the data. This allows the consumer application to parse individual data items when reading the data from the destination.

The PutRecord operation returns a RecordId, which is a unique string assigned to each record. Producer applications can use this ID for purposes such as auditability and investigation.

If the PutRecord operation throws a ServiceUnavailableException, back off and retry. If the exception persists, it is possible that the throughput limits have been exceeded for the delivery stream.

Data records sent to Kinesis Data Firehose are stored for 24 hours from the time they are added to a delivery stream as it tries to send the records to the destination. If the destination is unreachable for more than 24 hours, the data is no longer available.

Don't concatenate two or more base64 strings to form the data fields of your records. Instead, concatenate the raw data, then perform base64 encoding.

Creating a Request

data PutRecord

See: newPutRecord smart constructor.

Constructors

PutRecord'

Fields

deliveryStreamName :: Text

The name of the delivery stream.

record :: Record

The record.

newPutRecord

Arguments

:: Text
$sel:deliveryStreamName:PutRecord'

-> Record
$sel:record:PutRecord'

-> PutRecord

Create a value of PutRecord with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deliveryStreamName:PutRecord', putRecord_deliveryStreamName - The name of the delivery stream.

$sel:record:PutRecord', putRecord_record - The record.

Request Lenses

putRecord_deliveryStreamName :: Lens' PutRecord Text

The name of the delivery stream.

putRecord_record :: Lens' PutRecord Record

The record.

Destructuring the Response

data PutRecordResponse

See: newPutRecordResponse smart constructor.

Constructors

PutRecordResponse'

Fields

encrypted :: Maybe Bool

Indicates whether server-side encryption (SSE) was enabled during this operation.

httpStatus :: Int

The response's http status code.

recordId :: Text

The ID of the record.

newPutRecordResponse

Arguments

:: Int
$sel:httpStatus:PutRecordResponse'

-> Text
$sel:recordId:PutRecordResponse'

-> PutRecordResponse

Create a value of PutRecordResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:encrypted:PutRecordResponse', putRecordResponse_encrypted - Indicates whether server-side encryption (SSE) was enabled during this operation.

$sel:httpStatus:PutRecordResponse', putRecordResponse_httpStatus - The response's http status code.

$sel:recordId:PutRecordResponse', putRecordResponse_recordId - The ID of the record.

Response Lenses

putRecordResponse_encrypted :: Lens' PutRecordResponse (Maybe Bool)

Indicates whether server-side encryption (SSE) was enabled during this operation.

putRecordResponse_httpStatus :: Lens' PutRecordResponse Int

The response's http status code.

putRecordResponse_recordId :: Lens' PutRecordResponse Text

The ID of the record.