Network.AWS.Firehose.TagDeliveryStream

Description

Adds or updates tags for the specified delivery stream. A tag is a key-value pair that you can define and assign to AWS resources. If you specify a tag that already exists, the tag value is replaced with the value that you specify in the request. Tags are metadata. For example, you can add friendly names and descriptions or other types of information that can help you distinguish the delivery stream. For more information about tags, see Using Cost Allocation Tags in the AWS Billing and Cost Management User Guide.

Each delivery stream can have up to 50 tags.

This operation has a limit of five transactions per second per account.

Creating a Request

data TagDeliveryStream

See: newTagDeliveryStream smart constructor.

Constructors

TagDeliveryStream'

Fields

deliveryStreamName :: Text

The name of the delivery stream to which you want to add the tags.

tags :: NonEmpty Tag

A set of key-value pairs to use to create the tags.

newTagDeliveryStream

Arguments

:: Text
$sel:deliveryStreamName:TagDeliveryStream'

-> NonEmpty Tag
$sel:tags:TagDeliveryStream'

-> TagDeliveryStream

Create a value of TagDeliveryStream with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deliveryStreamName:TagDeliveryStream', tagDeliveryStream_deliveryStreamName - The name of the delivery stream to which you want to add the tags.

$sel:tags:TagDeliveryStream', tagDeliveryStream_tags - A set of key-value pairs to use to create the tags.

Request Lenses

tagDeliveryStream_deliveryStreamName :: Lens' TagDeliveryStream Text

The name of the delivery stream to which you want to add the tags.

tagDeliveryStream_tags :: Lens' TagDeliveryStream (NonEmpty Tag)

A set of key-value pairs to use to create the tags.

Destructuring the Response

data TagDeliveryStreamResponse

See: newTagDeliveryStreamResponse smart constructor.

Constructors

TagDeliveryStreamResponse'

Fields

httpStatus :: Int

The response's http status code.

newTagDeliveryStreamResponse

Arguments

:: Int
$sel:httpStatus:TagDeliveryStreamResponse'

-> TagDeliveryStreamResponse

Create a value of TagDeliveryStreamResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:TagDeliveryStreamResponse', tagDeliveryStreamResponse_httpStatus - The response's http status code.

Response Lenses

tagDeliveryStreamResponse_httpStatus :: Lens' TagDeliveryStreamResponse Int

The response's http status code.