Network.AWS.Firehose.Types.DeliveryStreamEncryptionConfiguration

Description

Documentation

data DeliveryStreamEncryptionConfiguration

Contains information about the server-side encryption (SSE) status for the delivery stream, the type customer master key (CMK) in use, if any, and the ARN of the CMK. You can get DeliveryStreamEncryptionConfiguration by invoking the DescribeDeliveryStream operation.

See: newDeliveryStreamEncryptionConfiguration smart constructor.

Constructors

DeliveryStreamEncryptionConfiguration'

Fields

status :: Maybe DeliveryStreamEncryptionStatus

This is the server-side encryption (SSE) status for the delivery stream. For a full description of the different values of this status, see StartDeliveryStreamEncryption and StopDeliveryStreamEncryption. If this status is ENABLING_FAILED or DISABLING_FAILED, it is the status of the most recent attempt to enable or disable SSE, respectively.

keyARN :: Maybe Text

If KeyType is CUSTOMER_MANAGED_CMK, this field contains the ARN of the customer managed CMK. If KeyType is AWS_OWNED_CMK, DeliveryStreamEncryptionConfiguration doesn't contain a value for KeyARN.

keyType :: Maybe KeyType

Indicates the type of customer master key (CMK) that is used for encryption. The default setting is AWS_OWNED_CMK. For more information about CMKs, see Customer Master Keys (CMKs).

failureDescription :: Maybe FailureDescription

Provides details in case one of the following operations fails due to an error related to KMS: CreateDeliveryStream, DeleteDeliveryStream, StartDeliveryStreamEncryption, StopDeliveryStreamEncryption.

newDeliveryStreamEncryptionConfiguration :: DeliveryStreamEncryptionConfiguration

Create a value of DeliveryStreamEncryptionConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DeliveryStreamEncryptionConfiguration', deliveryStreamEncryptionConfiguration_status - This is the server-side encryption (SSE) status for the delivery stream. For a full description of the different values of this status, see StartDeliveryStreamEncryption and StopDeliveryStreamEncryption. If this status is ENABLING_FAILED or DISABLING_FAILED, it is the status of the most recent attempt to enable or disable SSE, respectively.

$sel:keyARN:DeliveryStreamEncryptionConfiguration', deliveryStreamEncryptionConfiguration_keyARN - If KeyType is CUSTOMER_MANAGED_CMK, this field contains the ARN of the customer managed CMK. If KeyType is AWS_OWNED_CMK, DeliveryStreamEncryptionConfiguration doesn't contain a value for KeyARN.

$sel:keyType:DeliveryStreamEncryptionConfiguration', deliveryStreamEncryptionConfiguration_keyType - Indicates the type of customer master key (CMK) that is used for encryption. The default setting is AWS_OWNED_CMK. For more information about CMKs, see Customer Master Keys (CMKs).

$sel:failureDescription:DeliveryStreamEncryptionConfiguration', deliveryStreamEncryptionConfiguration_failureDescription - Provides details in case one of the following operations fails due to an error related to KMS: CreateDeliveryStream, DeleteDeliveryStream, StartDeliveryStreamEncryption, StopDeliveryStreamEncryption.

deliveryStreamEncryptionConfiguration_status :: Lens' DeliveryStreamEncryptionConfiguration (Maybe DeliveryStreamEncryptionStatus)

This is the server-side encryption (SSE) status for the delivery stream. For a full description of the different values of this status, see StartDeliveryStreamEncryption and StopDeliveryStreamEncryption. If this status is ENABLING_FAILED or DISABLING_FAILED, it is the status of the most recent attempt to enable or disable SSE, respectively.

deliveryStreamEncryptionConfiguration_keyARN :: Lens' DeliveryStreamEncryptionConfiguration (Maybe Text)

If KeyType is CUSTOMER_MANAGED_CMK, this field contains the ARN of the customer managed CMK. If KeyType is AWS_OWNED_CMK, DeliveryStreamEncryptionConfiguration doesn't contain a value for KeyARN.

deliveryStreamEncryptionConfiguration_keyType :: Lens' DeliveryStreamEncryptionConfiguration (Maybe KeyType)

Indicates the type of customer master key (CMK) that is used for encryption. The default setting is AWS_OWNED_CMK. For more information about CMKs, see Customer Master Keys (CMKs).

deliveryStreamEncryptionConfiguration_failureDescription :: Lens' DeliveryStreamEncryptionConfiguration (Maybe FailureDescription)

Provides details in case one of the following operations fails due to an error related to KMS: CreateDeliveryStream, DeleteDeliveryStream, StartDeliveryStreamEncryption, StopDeliveryStreamEncryption.