Network.AWS.Firehose.Types.ElasticsearchDestinationConfiguration

Description

Documentation

data ElasticsearchDestinationConfiguration

Describes the configuration of a destination in Amazon ES.

See: newElasticsearchDestinationConfiguration smart constructor.

Constructors

ElasticsearchDestinationConfiguration'

Fields

typeName :: Maybe Text

The Elasticsearch type name. For Elasticsearch 6.x, there can be only one type per index. If you try to specify a new type for an existing index that already has another type, Kinesis Data Firehose returns an error during run time.

For Elasticsearch 7.x, don't specify a TypeName.

clusterEndpoint :: Maybe Text

The endpoint to use when communicating with the cluster. Specify either this ClusterEndpoint or the DomainARN field.

processingConfiguration :: Maybe ProcessingConfiguration

The data processing configuration.

cloudWatchLoggingOptions :: Maybe CloudWatchLoggingOptions

The Amazon CloudWatch logging options for your delivery stream.

domainARN :: Maybe Text

The ARN of the Amazon ES domain. The IAM role must have permissions for DescribeElasticsearchDomain, DescribeElasticsearchDomains, and DescribeElasticsearchDomainConfig after assuming the role specified in RoleARN. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

Specify either ClusterEndpoint or DomainARN.

vpcConfiguration :: Maybe VpcConfiguration

The details of the VPC of the Amazon ES destination.

indexRotationPeriod :: Maybe ElasticsearchIndexRotationPeriod

The Elasticsearch index rotation period. Index rotation appends a timestamp to the IndexName to facilitate the expiration of old data. For more information, see Index Rotation for the Amazon ES Destination. The default value is OneDay.

bufferingHints :: Maybe ElasticsearchBufferingHints

The buffering options. If no value is specified, the default values for ElasticsearchBufferingHints are used.

retryOptions :: Maybe ElasticsearchRetryOptions

The retry behavior in case Kinesis Data Firehose is unable to deliver documents to Amazon ES. The default value is 300 (5 minutes).

s3BackupMode :: Maybe ElasticsearchS3BackupMode

Defines how documents should be delivered to Amazon S3. When it is set to FailedDocumentsOnly, Kinesis Data Firehose writes any documents that could not be indexed to the configured Amazon S3 destination, with elasticsearch-failed/ appended to the key prefix. When set to AllDocuments, Kinesis Data Firehose delivers all incoming records to Amazon S3, and also writes failed documents with elasticsearch-failed/ appended to the prefix. For more information, see Amazon S3 Backup for the Amazon ES Destination. Default value is FailedDocumentsOnly.

You can't change this backup mode after you create the delivery stream.

roleARN :: Text

The Amazon Resource Name (ARN) of the IAM role to be assumed by Kinesis Data Firehose for calling the Amazon ES Configuration API and for indexing documents. For more information, see Grant Kinesis Data Firehose Access to an Amazon S3 Destination and Amazon Resource Names (ARNs) and AWS Service Namespaces.

indexName :: Text

The Elasticsearch index name.

s3Configuration :: S3DestinationConfiguration

The configuration for the backup Amazon S3 location.

newElasticsearchDestinationConfiguration

Arguments

:: Text
$sel:roleARN:ElasticsearchDestinationConfiguration'

-> Text
$sel:indexName:ElasticsearchDestinationConfiguration'

-> S3DestinationConfiguration
$sel:s3Configuration:ElasticsearchDestinationConfiguration'

-> ElasticsearchDestinationConfiguration

Create a value of ElasticsearchDestinationConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:typeName:ElasticsearchDestinationConfiguration', elasticsearchDestinationConfiguration_typeName - The Elasticsearch type name. For Elasticsearch 6.x, there can be only one type per index. If you try to specify a new type for an existing index that already has another type, Kinesis Data Firehose returns an error during run time.

For Elasticsearch 7.x, don't specify a TypeName.

$sel:clusterEndpoint:ElasticsearchDestinationConfiguration', elasticsearchDestinationConfiguration_clusterEndpoint - The endpoint to use when communicating with the cluster. Specify either this ClusterEndpoint or the DomainARN field.

$sel:processingConfiguration:ElasticsearchDestinationConfiguration', elasticsearchDestinationConfiguration_processingConfiguration - The data processing configuration.

$sel:cloudWatchLoggingOptions:ElasticsearchDestinationConfiguration', elasticsearchDestinationConfiguration_cloudWatchLoggingOptions - The Amazon CloudWatch logging options for your delivery stream.

$sel:domainARN:ElasticsearchDestinationConfiguration', elasticsearchDestinationConfiguration_domainARN - The ARN of the Amazon ES domain. The IAM role must have permissions for DescribeElasticsearchDomain, DescribeElasticsearchDomains, and DescribeElasticsearchDomainConfig after assuming the role specified in RoleARN. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

Specify either ClusterEndpoint or DomainARN.

$sel:vpcConfiguration:ElasticsearchDestinationConfiguration', elasticsearchDestinationConfiguration_vpcConfiguration - The details of the VPC of the Amazon ES destination.

$sel:indexRotationPeriod:ElasticsearchDestinationConfiguration', elasticsearchDestinationConfiguration_indexRotationPeriod - The Elasticsearch index rotation period. Index rotation appends a timestamp to the IndexName to facilitate the expiration of old data. For more information, see Index Rotation for the Amazon ES Destination. The default value is OneDay.

$sel:bufferingHints:ElasticsearchDestinationConfiguration', elasticsearchDestinationConfiguration_bufferingHints - The buffering options. If no value is specified, the default values for ElasticsearchBufferingHints are used.

$sel:retryOptions:ElasticsearchDestinationConfiguration', elasticsearchDestinationConfiguration_retryOptions - The retry behavior in case Kinesis Data Firehose is unable to deliver documents to Amazon ES. The default value is 300 (5 minutes).

$sel:s3BackupMode:ElasticsearchDestinationConfiguration', elasticsearchDestinationConfiguration_s3BackupMode - Defines how documents should be delivered to Amazon S3. When it is set to FailedDocumentsOnly, Kinesis Data Firehose writes any documents that could not be indexed to the configured Amazon S3 destination, with elasticsearch-failed/ appended to the key prefix. When set to AllDocuments, Kinesis Data Firehose delivers all incoming records to Amazon S3, and also writes failed documents with elasticsearch-failed/ appended to the prefix. For more information, see Amazon S3 Backup for the Amazon ES Destination. Default value is FailedDocumentsOnly.

You can't change this backup mode after you create the delivery stream.

$sel:roleARN:ElasticsearchDestinationConfiguration', elasticsearchDestinationConfiguration_roleARN - The Amazon Resource Name (ARN) of the IAM role to be assumed by Kinesis Data Firehose for calling the Amazon ES Configuration API and for indexing documents. For more information, see Grant Kinesis Data Firehose Access to an Amazon S3 Destination and Amazon Resource Names (ARNs) and AWS Service Namespaces.

$sel:indexName:ElasticsearchDestinationConfiguration', elasticsearchDestinationConfiguration_indexName - The Elasticsearch index name.

$sel:s3Configuration:ElasticsearchDestinationConfiguration', elasticsearchDestinationConfiguration_s3Configuration - The configuration for the backup Amazon S3 location.

elasticsearchDestinationConfiguration_typeName :: Lens' ElasticsearchDestinationConfiguration (Maybe Text)

The Elasticsearch type name. For Elasticsearch 6.x, there can be only one type per index. If you try to specify a new type for an existing index that already has another type, Kinesis Data Firehose returns an error during run time.

For Elasticsearch 7.x, don't specify a TypeName.

elasticsearchDestinationConfiguration_clusterEndpoint :: Lens' ElasticsearchDestinationConfiguration (Maybe Text)

The endpoint to use when communicating with the cluster. Specify either this ClusterEndpoint or the DomainARN field.

elasticsearchDestinationConfiguration_processingConfiguration :: Lens' ElasticsearchDestinationConfiguration (Maybe ProcessingConfiguration)

The data processing configuration.

elasticsearchDestinationConfiguration_cloudWatchLoggingOptions :: Lens' ElasticsearchDestinationConfiguration (Maybe CloudWatchLoggingOptions)

The Amazon CloudWatch logging options for your delivery stream.

elasticsearchDestinationConfiguration_domainARN :: Lens' ElasticsearchDestinationConfiguration (Maybe Text)

The ARN of the Amazon ES domain. The IAM role must have permissions for DescribeElasticsearchDomain, DescribeElasticsearchDomains, and DescribeElasticsearchDomainConfig after assuming the role specified in RoleARN. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

Specify either ClusterEndpoint or DomainARN.

elasticsearchDestinationConfiguration_vpcConfiguration :: Lens' ElasticsearchDestinationConfiguration (Maybe VpcConfiguration)

The details of the VPC of the Amazon ES destination.

elasticsearchDestinationConfiguration_indexRotationPeriod :: Lens' ElasticsearchDestinationConfiguration (Maybe ElasticsearchIndexRotationPeriod)

The Elasticsearch index rotation period. Index rotation appends a timestamp to the IndexName to facilitate the expiration of old data. For more information, see Index Rotation for the Amazon ES Destination. The default value is OneDay.

elasticsearchDestinationConfiguration_bufferingHints :: Lens' ElasticsearchDestinationConfiguration (Maybe ElasticsearchBufferingHints)

The buffering options. If no value is specified, the default values for ElasticsearchBufferingHints are used.

elasticsearchDestinationConfiguration_retryOptions :: Lens' ElasticsearchDestinationConfiguration (Maybe ElasticsearchRetryOptions)

The retry behavior in case Kinesis Data Firehose is unable to deliver documents to Amazon ES. The default value is 300 (5 minutes).

elasticsearchDestinationConfiguration_s3BackupMode :: Lens' ElasticsearchDestinationConfiguration (Maybe ElasticsearchS3BackupMode)

Defines how documents should be delivered to Amazon S3. When it is set to FailedDocumentsOnly, Kinesis Data Firehose writes any documents that could not be indexed to the configured Amazon S3 destination, with elasticsearch-failed/ appended to the key prefix. When set to AllDocuments, Kinesis Data Firehose delivers all incoming records to Amazon S3, and also writes failed documents with elasticsearch-failed/ appended to the prefix. For more information, see Amazon S3 Backup for the Amazon ES Destination. Default value is FailedDocumentsOnly.

You can't change this backup mode after you create the delivery stream.

elasticsearchDestinationConfiguration_roleARN :: Lens' ElasticsearchDestinationConfiguration Text

The Amazon Resource Name (ARN) of the IAM role to be assumed by Kinesis Data Firehose for calling the Amazon ES Configuration API and for indexing documents. For more information, see Grant Kinesis Data Firehose Access to an Amazon S3 Destination and Amazon Resource Names (ARNs) and AWS Service Namespaces.

elasticsearchDestinationConfiguration_indexName :: Lens' ElasticsearchDestinationConfiguration Text

The Elasticsearch index name.

elasticsearchDestinationConfiguration_s3Configuration :: Lens' ElasticsearchDestinationConfiguration S3DestinationConfiguration

The configuration for the backup Amazon S3 location.