Network.AWS.Firehose.Types.ElasticsearchDestinationDescription

Description

Documentation

data ElasticsearchDestinationDescription

The destination description in Amazon ES.

See: newElasticsearchDestinationDescription smart constructor.

Constructors

ElasticsearchDestinationDescription'

Fields

indexRotationPeriod :: Maybe ElasticsearchIndexRotationPeriod

The Elasticsearch index rotation period

typeName :: Maybe Text

The Elasticsearch type name. This applies to Elasticsearch 6.x and lower versions. For Elasticsearch 7.x, there's no value for TypeName.

s3BackupMode :: Maybe ElasticsearchS3BackupMode

The Amazon S3 backup mode.

domainARN :: Maybe Text

The ARN of the Amazon ES domain. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

Kinesis Data Firehose uses either ClusterEndpoint or DomainARN to send data to Amazon ES.

vpcConfigurationDescription :: Maybe VpcConfigurationDescription

The details of the VPC of the Amazon ES destination.

cloudWatchLoggingOptions :: Maybe CloudWatchLoggingOptions

The Amazon CloudWatch logging options.

s3DestinationDescription :: Maybe S3DestinationDescription

The Amazon S3 destination.

bufferingHints :: Maybe ElasticsearchBufferingHints

The buffering options.

retryOptions :: Maybe ElasticsearchRetryOptions

The Amazon ES retry options.

processingConfiguration :: Maybe ProcessingConfiguration

The data processing configuration.

roleARN :: Maybe Text

The Amazon Resource Name (ARN) of the AWS credentials. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

clusterEndpoint :: Maybe Text

The endpoint to use when communicating with the cluster. Kinesis Data Firehose uses either this ClusterEndpoint or the DomainARN field to send data to Amazon ES.

indexName :: Maybe Text

The Elasticsearch index name.

newElasticsearchDestinationDescription :: ElasticsearchDestinationDescription

Create a value of ElasticsearchDestinationDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:indexRotationPeriod:ElasticsearchDestinationDescription', elasticsearchDestinationDescription_indexRotationPeriod - The Elasticsearch index rotation period

$sel:typeName:ElasticsearchDestinationDescription', elasticsearchDestinationDescription_typeName - The Elasticsearch type name. This applies to Elasticsearch 6.x and lower versions. For Elasticsearch 7.x, there's no value for TypeName.

$sel:s3BackupMode:ElasticsearchDestinationDescription', elasticsearchDestinationDescription_s3BackupMode - The Amazon S3 backup mode.

$sel:domainARN:ElasticsearchDestinationDescription', elasticsearchDestinationDescription_domainARN - The ARN of the Amazon ES domain. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

Kinesis Data Firehose uses either ClusterEndpoint or DomainARN to send data to Amazon ES.

$sel:vpcConfigurationDescription:ElasticsearchDestinationDescription', elasticsearchDestinationDescription_vpcConfigurationDescription - The details of the VPC of the Amazon ES destination.

$sel:cloudWatchLoggingOptions:ElasticsearchDestinationDescription', elasticsearchDestinationDescription_cloudWatchLoggingOptions - The Amazon CloudWatch logging options.

$sel:s3DestinationDescription:ElasticsearchDestinationDescription', elasticsearchDestinationDescription_s3DestinationDescription - The Amazon S3 destination.

$sel:bufferingHints:ElasticsearchDestinationDescription', elasticsearchDestinationDescription_bufferingHints - The buffering options.

$sel:retryOptions:ElasticsearchDestinationDescription', elasticsearchDestinationDescription_retryOptions - The Amazon ES retry options.

$sel:processingConfiguration:ElasticsearchDestinationDescription', elasticsearchDestinationDescription_processingConfiguration - The data processing configuration.

$sel:roleARN:ElasticsearchDestinationDescription', elasticsearchDestinationDescription_roleARN - The Amazon Resource Name (ARN) of the AWS credentials. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

$sel:clusterEndpoint:ElasticsearchDestinationDescription', elasticsearchDestinationDescription_clusterEndpoint - The endpoint to use when communicating with the cluster. Kinesis Data Firehose uses either this ClusterEndpoint or the DomainARN field to send data to Amazon ES.

$sel:indexName:ElasticsearchDestinationDescription', elasticsearchDestinationDescription_indexName - The Elasticsearch index name.

elasticsearchDestinationDescription_indexRotationPeriod :: Lens' ElasticsearchDestinationDescription (Maybe ElasticsearchIndexRotationPeriod)

The Elasticsearch index rotation period

elasticsearchDestinationDescription_typeName :: Lens' ElasticsearchDestinationDescription (Maybe Text)

The Elasticsearch type name. This applies to Elasticsearch 6.x and lower versions. For Elasticsearch 7.x, there's no value for TypeName.

elasticsearchDestinationDescription_s3BackupMode :: Lens' ElasticsearchDestinationDescription (Maybe ElasticsearchS3BackupMode)

The Amazon S3 backup mode.

elasticsearchDestinationDescription_domainARN :: Lens' ElasticsearchDestinationDescription (Maybe Text)

The ARN of the Amazon ES domain. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

Kinesis Data Firehose uses either ClusterEndpoint or DomainARN to send data to Amazon ES.

elasticsearchDestinationDescription_vpcConfigurationDescription :: Lens' ElasticsearchDestinationDescription (Maybe VpcConfigurationDescription)

The details of the VPC of the Amazon ES destination.

elasticsearchDestinationDescription_cloudWatchLoggingOptions :: Lens' ElasticsearchDestinationDescription (Maybe CloudWatchLoggingOptions)

The Amazon CloudWatch logging options.

elasticsearchDestinationDescription_s3DestinationDescription :: Lens' ElasticsearchDestinationDescription (Maybe S3DestinationDescription)

The Amazon S3 destination.

elasticsearchDestinationDescription_bufferingHints :: Lens' ElasticsearchDestinationDescription (Maybe ElasticsearchBufferingHints)

The buffering options.

elasticsearchDestinationDescription_retryOptions :: Lens' ElasticsearchDestinationDescription (Maybe ElasticsearchRetryOptions)

The Amazon ES retry options.

elasticsearchDestinationDescription_processingConfiguration :: Lens' ElasticsearchDestinationDescription (Maybe ProcessingConfiguration)

The data processing configuration.

elasticsearchDestinationDescription_roleARN :: Lens' ElasticsearchDestinationDescription (Maybe Text)

The Amazon Resource Name (ARN) of the AWS credentials. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

elasticsearchDestinationDescription_clusterEndpoint :: Lens' ElasticsearchDestinationDescription (Maybe Text)

The endpoint to use when communicating with the cluster. Kinesis Data Firehose uses either this ClusterEndpoint or the DomainARN field to send data to Amazon ES.

elasticsearchDestinationDescription_indexName :: Lens' ElasticsearchDestinationDescription (Maybe Text)

The Elasticsearch index name.