Network.AWS.Firehose.Types.ElasticsearchDestinationUpdate

Description

Documentation

data ElasticsearchDestinationUpdate

Describes an update for a destination in Amazon ES.

See: newElasticsearchDestinationUpdate smart constructor.

Constructors

ElasticsearchDestinationUpdate'

Fields

typeName :: Maybe Text

The Elasticsearch type name. For Elasticsearch 6.x, there can be only one type per index. If you try to specify a new type for an existing index that already has another type, Kinesis Data Firehose returns an error during runtime.

If you upgrade Elasticsearch from 6.x to 7.x and don’t update your delivery stream, Kinesis Data Firehose still delivers data to Elasticsearch with the old index name and type name. If you want to update your delivery stream with a new index name, provide an empty string for TypeName.

roleARN :: Maybe Text

The Amazon Resource Name (ARN) of the IAM role to be assumed by Kinesis Data Firehose for calling the Amazon ES Configuration API and for indexing documents. For more information, see Grant Kinesis Data Firehose Access to an Amazon S3 Destination and Amazon Resource Names (ARNs) and AWS Service Namespaces.

clusterEndpoint :: Maybe Text

The endpoint to use when communicating with the cluster. Specify either this ClusterEndpoint or the DomainARN field.

indexName :: Maybe Text

The Elasticsearch index name.

s3Update :: Maybe S3DestinationUpdate

The Amazon S3 destination.

processingConfiguration :: Maybe ProcessingConfiguration

The data processing configuration.

cloudWatchLoggingOptions :: Maybe CloudWatchLoggingOptions

The CloudWatch logging options for your delivery stream.

domainARN :: Maybe Text

The ARN of the Amazon ES domain. The IAM role must have permissions for DescribeElasticsearchDomain, DescribeElasticsearchDomains, and DescribeElasticsearchDomainConfig after assuming the IAM role specified in RoleARN. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

Specify either ClusterEndpoint or DomainARN.

indexRotationPeriod :: Maybe ElasticsearchIndexRotationPeriod

The Elasticsearch index rotation period. Index rotation appends a timestamp to IndexName to facilitate the expiration of old data. For more information, see Index Rotation for the Amazon ES Destination. Default value is OneDay.

bufferingHints :: Maybe ElasticsearchBufferingHints

The buffering options. If no value is specified, ElasticsearchBufferingHints object default values are used.

retryOptions :: Maybe ElasticsearchRetryOptions

The retry behavior in case Kinesis Data Firehose is unable to deliver documents to Amazon ES. The default value is 300 (5 minutes).

newElasticsearchDestinationUpdate :: ElasticsearchDestinationUpdate

Create a value of ElasticsearchDestinationUpdate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:typeName:ElasticsearchDestinationUpdate', elasticsearchDestinationUpdate_typeName - The Elasticsearch type name. For Elasticsearch 6.x, there can be only one type per index. If you try to specify a new type for an existing index that already has another type, Kinesis Data Firehose returns an error during runtime.

If you upgrade Elasticsearch from 6.x to 7.x and don’t update your delivery stream, Kinesis Data Firehose still delivers data to Elasticsearch with the old index name and type name. If you want to update your delivery stream with a new index name, provide an empty string for TypeName.

$sel:roleARN:ElasticsearchDestinationUpdate', elasticsearchDestinationUpdate_roleARN - The Amazon Resource Name (ARN) of the IAM role to be assumed by Kinesis Data Firehose for calling the Amazon ES Configuration API and for indexing documents. For more information, see Grant Kinesis Data Firehose Access to an Amazon S3 Destination and Amazon Resource Names (ARNs) and AWS Service Namespaces.

$sel:clusterEndpoint:ElasticsearchDestinationUpdate', elasticsearchDestinationUpdate_clusterEndpoint - The endpoint to use when communicating with the cluster. Specify either this ClusterEndpoint or the DomainARN field.

$sel:indexName:ElasticsearchDestinationUpdate', elasticsearchDestinationUpdate_indexName - The Elasticsearch index name.

$sel:s3Update:ElasticsearchDestinationUpdate', elasticsearchDestinationUpdate_s3Update - The Amazon S3 destination.

$sel:processingConfiguration:ElasticsearchDestinationUpdate', elasticsearchDestinationUpdate_processingConfiguration - The data processing configuration.

$sel:cloudWatchLoggingOptions:ElasticsearchDestinationUpdate', elasticsearchDestinationUpdate_cloudWatchLoggingOptions - The CloudWatch logging options for your delivery stream.

$sel:domainARN:ElasticsearchDestinationUpdate', elasticsearchDestinationUpdate_domainARN - The ARN of the Amazon ES domain. The IAM role must have permissions for DescribeElasticsearchDomain, DescribeElasticsearchDomains, and DescribeElasticsearchDomainConfig after assuming the IAM role specified in RoleARN. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

Specify either ClusterEndpoint or DomainARN.

$sel:indexRotationPeriod:ElasticsearchDestinationUpdate', elasticsearchDestinationUpdate_indexRotationPeriod - The Elasticsearch index rotation period. Index rotation appends a timestamp to IndexName to facilitate the expiration of old data. For more information, see Index Rotation for the Amazon ES Destination. Default value is OneDay.

$sel:bufferingHints:ElasticsearchDestinationUpdate', elasticsearchDestinationUpdate_bufferingHints - The buffering options. If no value is specified, ElasticsearchBufferingHints object default values are used.

$sel:retryOptions:ElasticsearchDestinationUpdate', elasticsearchDestinationUpdate_retryOptions - The retry behavior in case Kinesis Data Firehose is unable to deliver documents to Amazon ES. The default value is 300 (5 minutes).

elasticsearchDestinationUpdate_typeName :: Lens' ElasticsearchDestinationUpdate (Maybe Text)

The Elasticsearch type name. For Elasticsearch 6.x, there can be only one type per index. If you try to specify a new type for an existing index that already has another type, Kinesis Data Firehose returns an error during runtime.

If you upgrade Elasticsearch from 6.x to 7.x and don’t update your delivery stream, Kinesis Data Firehose still delivers data to Elasticsearch with the old index name and type name. If you want to update your delivery stream with a new index name, provide an empty string for TypeName.

elasticsearchDestinationUpdate_roleARN :: Lens' ElasticsearchDestinationUpdate (Maybe Text)

The Amazon Resource Name (ARN) of the IAM role to be assumed by Kinesis Data Firehose for calling the Amazon ES Configuration API and for indexing documents. For more information, see Grant Kinesis Data Firehose Access to an Amazon S3 Destination and Amazon Resource Names (ARNs) and AWS Service Namespaces.

elasticsearchDestinationUpdate_clusterEndpoint :: Lens' ElasticsearchDestinationUpdate (Maybe Text)

The endpoint to use when communicating with the cluster. Specify either this ClusterEndpoint or the DomainARN field.

elasticsearchDestinationUpdate_indexName :: Lens' ElasticsearchDestinationUpdate (Maybe Text)

The Elasticsearch index name.

elasticsearchDestinationUpdate_s3Update :: Lens' ElasticsearchDestinationUpdate (Maybe S3DestinationUpdate)

The Amazon S3 destination.

elasticsearchDestinationUpdate_processingConfiguration :: Lens' ElasticsearchDestinationUpdate (Maybe ProcessingConfiguration)

The data processing configuration.

elasticsearchDestinationUpdate_cloudWatchLoggingOptions :: Lens' ElasticsearchDestinationUpdate (Maybe CloudWatchLoggingOptions)

The CloudWatch logging options for your delivery stream.

elasticsearchDestinationUpdate_domainARN :: Lens' ElasticsearchDestinationUpdate (Maybe Text)

The ARN of the Amazon ES domain. The IAM role must have permissions for DescribeElasticsearchDomain, DescribeElasticsearchDomains, and DescribeElasticsearchDomainConfig after assuming the IAM role specified in RoleARN. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

Specify either ClusterEndpoint or DomainARN.

elasticsearchDestinationUpdate_indexRotationPeriod :: Lens' ElasticsearchDestinationUpdate (Maybe ElasticsearchIndexRotationPeriod)

The Elasticsearch index rotation period. Index rotation appends a timestamp to IndexName to facilitate the expiration of old data. For more information, see Index Rotation for the Amazon ES Destination. Default value is OneDay.

elasticsearchDestinationUpdate_bufferingHints :: Lens' ElasticsearchDestinationUpdate (Maybe ElasticsearchBufferingHints)

The buffering options. If no value is specified, ElasticsearchBufferingHints object default values are used.

elasticsearchDestinationUpdate_retryOptions :: Lens' ElasticsearchDestinationUpdate (Maybe ElasticsearchRetryOptions)

The retry behavior in case Kinesis Data Firehose is unable to deliver documents to Amazon ES. The default value is 300 (5 minutes).