Network.AWS.Firehose.Types.ExtendedS3DestinationConfiguration

Description

Documentation

data ExtendedS3DestinationConfiguration

Describes the configuration of a destination in Amazon S3.

See: newExtendedS3DestinationConfiguration smart constructor.

Constructors

ExtendedS3DestinationConfiguration'

Fields

errorOutputPrefix :: Maybe Text

A prefix that Kinesis Data Firehose evaluates and adds to failed records before writing them to S3. This prefix appears immediately following the bucket name. For information about how to specify this prefix, see Custom Prefixes for Amazon S3 Objects.

encryptionConfiguration :: Maybe EncryptionConfiguration

The encryption configuration. If no value is specified, the default is no encryption.

s3BackupConfiguration :: Maybe S3DestinationConfiguration

The configuration for backup in Amazon S3.

processingConfiguration :: Maybe ProcessingConfiguration

The data processing configuration.

cloudWatchLoggingOptions :: Maybe CloudWatchLoggingOptions

The Amazon CloudWatch logging options for your delivery stream.

dataFormatConversionConfiguration :: Maybe DataFormatConversionConfiguration

The serializer, deserializer, and schema for converting data from the JSON format to the Parquet or ORC format before writing it to Amazon S3.

prefix :: Maybe Text

The "YYYY/MM/DD/HH" time format prefix is automatically used for delivered Amazon S3 files. You can also specify a custom prefix, as described in Custom Prefixes for Amazon S3 Objects.

dynamicPartitioningConfiguration :: Maybe DynamicPartitioningConfiguration

The configuration of the dynamic partitioning mechanism that creates smaller data sets from the streaming data by partitioning it based on partition keys. Currently, dynamic partitioning is only supported for Amazon S3 destinations. For more information, see https://docs.aws.amazon.com/firehose/latest/dev/dynamic-partitioning.html

bufferingHints :: Maybe BufferingHints

The buffering option.

s3BackupMode :: Maybe S3BackupMode

The Amazon S3 backup mode. After you create a delivery stream, you can update it to enable Amazon S3 backup if it is disabled. If backup is enabled, you can't update the delivery stream to disable it.

compressionFormat :: Maybe CompressionFormat

The compression format. If no value is specified, the default is UNCOMPRESSED.

roleARN :: Text

The Amazon Resource Name (ARN) of the AWS credentials. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

bucketARN :: Text

The ARN of the S3 bucket. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

newExtendedS3DestinationConfiguration

Arguments

:: Text
$sel:roleARN:ExtendedS3DestinationConfiguration'

-> Text
$sel:bucketARN:ExtendedS3DestinationConfiguration'

-> ExtendedS3DestinationConfiguration

Create a value of ExtendedS3DestinationConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:errorOutputPrefix:ExtendedS3DestinationConfiguration', extendedS3DestinationConfiguration_errorOutputPrefix - A prefix that Kinesis Data Firehose evaluates and adds to failed records before writing them to S3. This prefix appears immediately following the bucket name. For information about how to specify this prefix, see Custom Prefixes for Amazon S3 Objects.

$sel:encryptionConfiguration:ExtendedS3DestinationConfiguration', extendedS3DestinationConfiguration_encryptionConfiguration - The encryption configuration. If no value is specified, the default is no encryption.

$sel:s3BackupConfiguration:ExtendedS3DestinationConfiguration', extendedS3DestinationConfiguration_s3BackupConfiguration - The configuration for backup in Amazon S3.

$sel:processingConfiguration:ExtendedS3DestinationConfiguration', extendedS3DestinationConfiguration_processingConfiguration - The data processing configuration.

$sel:cloudWatchLoggingOptions:ExtendedS3DestinationConfiguration', extendedS3DestinationConfiguration_cloudWatchLoggingOptions - The Amazon CloudWatch logging options for your delivery stream.

$sel:dataFormatConversionConfiguration:ExtendedS3DestinationConfiguration', extendedS3DestinationConfiguration_dataFormatConversionConfiguration - The serializer, deserializer, and schema for converting data from the JSON format to the Parquet or ORC format before writing it to Amazon S3.

$sel:prefix:ExtendedS3DestinationConfiguration', extendedS3DestinationConfiguration_prefix - The "YYYY/MM/DD/HH" time format prefix is automatically used for delivered Amazon S3 files. You can also specify a custom prefix, as described in Custom Prefixes for Amazon S3 Objects.

$sel:dynamicPartitioningConfiguration:ExtendedS3DestinationConfiguration', extendedS3DestinationConfiguration_dynamicPartitioningConfiguration - The configuration of the dynamic partitioning mechanism that creates smaller data sets from the streaming data by partitioning it based on partition keys. Currently, dynamic partitioning is only supported for Amazon S3 destinations. For more information, see https://docs.aws.amazon.com/firehose/latest/dev/dynamic-partitioning.html

$sel:bufferingHints:ExtendedS3DestinationConfiguration', extendedS3DestinationConfiguration_bufferingHints - The buffering option.

$sel:s3BackupMode:ExtendedS3DestinationConfiguration', extendedS3DestinationConfiguration_s3BackupMode - The Amazon S3 backup mode. After you create a delivery stream, you can update it to enable Amazon S3 backup if it is disabled. If backup is enabled, you can't update the delivery stream to disable it.

$sel:compressionFormat:ExtendedS3DestinationConfiguration', extendedS3DestinationConfiguration_compressionFormat - The compression format. If no value is specified, the default is UNCOMPRESSED.

$sel:roleARN:ExtendedS3DestinationConfiguration', extendedS3DestinationConfiguration_roleARN - The Amazon Resource Name (ARN) of the AWS credentials. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

$sel:bucketARN:ExtendedS3DestinationConfiguration', extendedS3DestinationConfiguration_bucketARN - The ARN of the S3 bucket. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

extendedS3DestinationConfiguration_errorOutputPrefix :: Lens' ExtendedS3DestinationConfiguration (Maybe Text)

A prefix that Kinesis Data Firehose evaluates and adds to failed records before writing them to S3. This prefix appears immediately following the bucket name. For information about how to specify this prefix, see Custom Prefixes for Amazon S3 Objects.

extendedS3DestinationConfiguration_encryptionConfiguration :: Lens' ExtendedS3DestinationConfiguration (Maybe EncryptionConfiguration)

The encryption configuration. If no value is specified, the default is no encryption.

extendedS3DestinationConfiguration_s3BackupConfiguration :: Lens' ExtendedS3DestinationConfiguration (Maybe S3DestinationConfiguration)

The configuration for backup in Amazon S3.

extendedS3DestinationConfiguration_processingConfiguration :: Lens' ExtendedS3DestinationConfiguration (Maybe ProcessingConfiguration)

The data processing configuration.

extendedS3DestinationConfiguration_cloudWatchLoggingOptions :: Lens' ExtendedS3DestinationConfiguration (Maybe CloudWatchLoggingOptions)

The Amazon CloudWatch logging options for your delivery stream.

extendedS3DestinationConfiguration_dataFormatConversionConfiguration :: Lens' ExtendedS3DestinationConfiguration (Maybe DataFormatConversionConfiguration)

The serializer, deserializer, and schema for converting data from the JSON format to the Parquet or ORC format before writing it to Amazon S3.

extendedS3DestinationConfiguration_prefix :: Lens' ExtendedS3DestinationConfiguration (Maybe Text)

The "YYYY/MM/DD/HH" time format prefix is automatically used for delivered Amazon S3 files. You can also specify a custom prefix, as described in Custom Prefixes for Amazon S3 Objects.

extendedS3DestinationConfiguration_dynamicPartitioningConfiguration :: Lens' ExtendedS3DestinationConfiguration (Maybe DynamicPartitioningConfiguration)

The configuration of the dynamic partitioning mechanism that creates smaller data sets from the streaming data by partitioning it based on partition keys. Currently, dynamic partitioning is only supported for Amazon S3 destinations. For more information, see https://docs.aws.amazon.com/firehose/latest/dev/dynamic-partitioning.html

extendedS3DestinationConfiguration_bufferingHints :: Lens' ExtendedS3DestinationConfiguration (Maybe BufferingHints)

The buffering option.

extendedS3DestinationConfiguration_s3BackupMode :: Lens' ExtendedS3DestinationConfiguration (Maybe S3BackupMode)

The Amazon S3 backup mode. After you create a delivery stream, you can update it to enable Amazon S3 backup if it is disabled. If backup is enabled, you can't update the delivery stream to disable it.

extendedS3DestinationConfiguration_compressionFormat :: Lens' ExtendedS3DestinationConfiguration (Maybe CompressionFormat)

The compression format. If no value is specified, the default is UNCOMPRESSED.

extendedS3DestinationConfiguration_roleARN :: Lens' ExtendedS3DestinationConfiguration Text

The Amazon Resource Name (ARN) of the AWS credentials. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

extendedS3DestinationConfiguration_bucketARN :: Lens' ExtendedS3DestinationConfiguration Text

The ARN of the S3 bucket. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces.