Network.AWS.Firehose.Types.ExtendedS3DestinationDescription

Description

Documentation

data ExtendedS3DestinationDescription

Describes a destination in Amazon S3.

See: newExtendedS3DestinationDescription smart constructor.

Constructors

ExtendedS3DestinationDescription'

Fields

s3BackupMode :: Maybe S3BackupMode

The Amazon S3 backup mode.

s3BackupDescription :: Maybe S3DestinationDescription

The configuration for backup in Amazon S3.

prefix :: Maybe Text

The "YYYY/MM/DD/HH" time format prefix is automatically used for delivered Amazon S3 files. You can also specify a custom prefix, as described in Custom Prefixes for Amazon S3 Objects.

cloudWatchLoggingOptions :: Maybe CloudWatchLoggingOptions

The Amazon CloudWatch logging options for your delivery stream.

errorOutputPrefix :: Maybe Text

A prefix that Kinesis Data Firehose evaluates and adds to failed records before writing them to S3. This prefix appears immediately following the bucket name. For information about how to specify this prefix, see Custom Prefixes for Amazon S3 Objects.

dynamicPartitioningConfiguration :: Maybe DynamicPartitioningConfiguration

The configuration of the dynamic partitioning mechanism that creates smaller data sets from the streaming data by partitioning it based on partition keys. Currently, dynamic partitioning is only supported for Amazon S3 destinations. For more information, see https://docs.aws.amazon.com/firehose/latest/dev/dynamic-partitioning.html

dataFormatConversionConfiguration :: Maybe DataFormatConversionConfiguration

The serializer, deserializer, and schema for converting data from the JSON format to the Parquet or ORC format before writing it to Amazon S3.

processingConfiguration :: Maybe ProcessingConfiguration

The data processing configuration.

roleARN :: Text

The Amazon Resource Name (ARN) of the AWS credentials. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

bucketARN :: Text

The ARN of the S3 bucket. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

bufferingHints :: BufferingHints

The buffering option.

compressionFormat :: CompressionFormat

The compression format. If no value is specified, the default is UNCOMPRESSED.

encryptionConfiguration :: EncryptionConfiguration

The encryption configuration. If no value is specified, the default is no encryption.

newExtendedS3DestinationDescription

Arguments

:: Text
$sel:roleARN:ExtendedS3DestinationDescription'

-> Text
$sel:bucketARN:ExtendedS3DestinationDescription'

-> BufferingHints
$sel:bufferingHints:ExtendedS3DestinationDescription'

-> CompressionFormat
$sel:compressionFormat:ExtendedS3DestinationDescription'

-> EncryptionConfiguration
$sel:encryptionConfiguration:ExtendedS3DestinationDescription'

-> ExtendedS3DestinationDescription

Create a value of ExtendedS3DestinationDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3BackupMode:ExtendedS3DestinationDescription', extendedS3DestinationDescription_s3BackupMode - The Amazon S3 backup mode.

$sel:s3BackupDescription:ExtendedS3DestinationDescription', extendedS3DestinationDescription_s3BackupDescription - The configuration for backup in Amazon S3.

$sel:prefix:ExtendedS3DestinationDescription', extendedS3DestinationDescription_prefix - The "YYYY/MM/DD/HH" time format prefix is automatically used for delivered Amazon S3 files. You can also specify a custom prefix, as described in Custom Prefixes for Amazon S3 Objects.

$sel:cloudWatchLoggingOptions:ExtendedS3DestinationDescription', extendedS3DestinationDescription_cloudWatchLoggingOptions - The Amazon CloudWatch logging options for your delivery stream.

$sel:errorOutputPrefix:ExtendedS3DestinationDescription', extendedS3DestinationDescription_errorOutputPrefix - A prefix that Kinesis Data Firehose evaluates and adds to failed records before writing them to S3. This prefix appears immediately following the bucket name. For information about how to specify this prefix, see Custom Prefixes for Amazon S3 Objects.

$sel:dynamicPartitioningConfiguration:ExtendedS3DestinationDescription', extendedS3DestinationDescription_dynamicPartitioningConfiguration - The configuration of the dynamic partitioning mechanism that creates smaller data sets from the streaming data by partitioning it based on partition keys. Currently, dynamic partitioning is only supported for Amazon S3 destinations. For more information, see https://docs.aws.amazon.com/firehose/latest/dev/dynamic-partitioning.html

$sel:dataFormatConversionConfiguration:ExtendedS3DestinationDescription', extendedS3DestinationDescription_dataFormatConversionConfiguration - The serializer, deserializer, and schema for converting data from the JSON format to the Parquet or ORC format before writing it to Amazon S3.

$sel:processingConfiguration:ExtendedS3DestinationDescription', extendedS3DestinationDescription_processingConfiguration - The data processing configuration.

$sel:roleARN:ExtendedS3DestinationDescription', extendedS3DestinationDescription_roleARN - The Amazon Resource Name (ARN) of the AWS credentials. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

$sel:bucketARN:ExtendedS3DestinationDescription', extendedS3DestinationDescription_bucketARN - The ARN of the S3 bucket. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

$sel:bufferingHints:ExtendedS3DestinationDescription', extendedS3DestinationDescription_bufferingHints - The buffering option.

$sel:compressionFormat:ExtendedS3DestinationDescription', extendedS3DestinationDescription_compressionFormat - The compression format. If no value is specified, the default is UNCOMPRESSED.

$sel:encryptionConfiguration:ExtendedS3DestinationDescription', extendedS3DestinationDescription_encryptionConfiguration - The encryption configuration. If no value is specified, the default is no encryption.

extendedS3DestinationDescription_s3BackupMode :: Lens' ExtendedS3DestinationDescription (Maybe S3BackupMode)

The Amazon S3 backup mode.

extendedS3DestinationDescription_s3BackupDescription :: Lens' ExtendedS3DestinationDescription (Maybe S3DestinationDescription)

The configuration for backup in Amazon S3.

extendedS3DestinationDescription_prefix :: Lens' ExtendedS3DestinationDescription (Maybe Text)

The "YYYY/MM/DD/HH" time format prefix is automatically used for delivered Amazon S3 files. You can also specify a custom prefix, as described in Custom Prefixes for Amazon S3 Objects.

extendedS3DestinationDescription_cloudWatchLoggingOptions :: Lens' ExtendedS3DestinationDescription (Maybe CloudWatchLoggingOptions)

The Amazon CloudWatch logging options for your delivery stream.

extendedS3DestinationDescription_errorOutputPrefix :: Lens' ExtendedS3DestinationDescription (Maybe Text)

A prefix that Kinesis Data Firehose evaluates and adds to failed records before writing them to S3. This prefix appears immediately following the bucket name. For information about how to specify this prefix, see Custom Prefixes for Amazon S3 Objects.

extendedS3DestinationDescription_dynamicPartitioningConfiguration :: Lens' ExtendedS3DestinationDescription (Maybe DynamicPartitioningConfiguration)

The configuration of the dynamic partitioning mechanism that creates smaller data sets from the streaming data by partitioning it based on partition keys. Currently, dynamic partitioning is only supported for Amazon S3 destinations. For more information, see https://docs.aws.amazon.com/firehose/latest/dev/dynamic-partitioning.html

extendedS3DestinationDescription_dataFormatConversionConfiguration :: Lens' ExtendedS3DestinationDescription (Maybe DataFormatConversionConfiguration)

The serializer, deserializer, and schema for converting data from the JSON format to the Parquet or ORC format before writing it to Amazon S3.

extendedS3DestinationDescription_processingConfiguration :: Lens' ExtendedS3DestinationDescription (Maybe ProcessingConfiguration)

The data processing configuration.

extendedS3DestinationDescription_roleARN :: Lens' ExtendedS3DestinationDescription Text

The Amazon Resource Name (ARN) of the AWS credentials. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

extendedS3DestinationDescription_bucketARN :: Lens' ExtendedS3DestinationDescription Text

The ARN of the S3 bucket. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

extendedS3DestinationDescription_bufferingHints :: Lens' ExtendedS3DestinationDescription BufferingHints

The buffering option.

extendedS3DestinationDescription_compressionFormat :: Lens' ExtendedS3DestinationDescription CompressionFormat

The compression format. If no value is specified, the default is UNCOMPRESSED.

extendedS3DestinationDescription_encryptionConfiguration :: Lens' ExtendedS3DestinationDescription EncryptionConfiguration

The encryption configuration. If no value is specified, the default is no encryption.