Network.AWS.Firehose.Types.ExtendedS3DestinationUpdate

Description

Documentation

data ExtendedS3DestinationUpdate

Describes an update for a destination in Amazon S3.

See: newExtendedS3DestinationUpdate smart constructor.

Constructors

ExtendedS3DestinationUpdate'

Fields

s3BackupMode :: Maybe S3BackupMode

You can update a delivery stream to enable Amazon S3 backup if it is disabled. If backup is enabled, you can't update the delivery stream to disable it.

prefix :: Maybe Text

The "YYYY/MM/DD/HH" time format prefix is automatically used for delivered Amazon S3 files. You can also specify a custom prefix, as described in Custom Prefixes for Amazon S3 Objects.

cloudWatchLoggingOptions :: Maybe CloudWatchLoggingOptions

The Amazon CloudWatch logging options for your delivery stream.

errorOutputPrefix :: Maybe Text

A prefix that Kinesis Data Firehose evaluates and adds to failed records before writing them to S3. This prefix appears immediately following the bucket name. For information about how to specify this prefix, see Custom Prefixes for Amazon S3 Objects.

s3BackupUpdate :: Maybe S3DestinationUpdate

The Amazon S3 destination for backup.

encryptionConfiguration :: Maybe EncryptionConfiguration

The encryption configuration. If no value is specified, the default is no encryption.

dynamicPartitioningConfiguration :: Maybe DynamicPartitioningConfiguration

The configuration of the dynamic partitioning mechanism that creates smaller data sets from the streaming data by partitioning it based on partition keys. Currently, dynamic partitioning is only supported for Amazon S3 destinations. For more information, see https://docs.aws.amazon.com/firehose/latest/dev/dynamic-partitioning.html

compressionFormat :: Maybe CompressionFormat

The compression format. If no value is specified, the default is UNCOMPRESSED.

bufferingHints :: Maybe BufferingHints

The buffering option.

dataFormatConversionConfiguration :: Maybe DataFormatConversionConfiguration

The serializer, deserializer, and schema for converting data from the JSON format to the Parquet or ORC format before writing it to Amazon S3.

bucketARN :: Maybe Text

The ARN of the S3 bucket. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

processingConfiguration :: Maybe ProcessingConfiguration

The data processing configuration.

roleARN :: Maybe Text

The Amazon Resource Name (ARN) of the AWS credentials. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

newExtendedS3DestinationUpdate :: ExtendedS3DestinationUpdate

Create a value of ExtendedS3DestinationUpdate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3BackupMode:ExtendedS3DestinationUpdate', extendedS3DestinationUpdate_s3BackupMode - You can update a delivery stream to enable Amazon S3 backup if it is disabled. If backup is enabled, you can't update the delivery stream to disable it.

$sel:prefix:ExtendedS3DestinationUpdate', extendedS3DestinationUpdate_prefix - The "YYYY/MM/DD/HH" time format prefix is automatically used for delivered Amazon S3 files. You can also specify a custom prefix, as described in Custom Prefixes for Amazon S3 Objects.

$sel:cloudWatchLoggingOptions:ExtendedS3DestinationUpdate', extendedS3DestinationUpdate_cloudWatchLoggingOptions - The Amazon CloudWatch logging options for your delivery stream.

$sel:errorOutputPrefix:ExtendedS3DestinationUpdate', extendedS3DestinationUpdate_errorOutputPrefix - A prefix that Kinesis Data Firehose evaluates and adds to failed records before writing them to S3. This prefix appears immediately following the bucket name. For information about how to specify this prefix, see Custom Prefixes for Amazon S3 Objects.

$sel:s3BackupUpdate:ExtendedS3DestinationUpdate', extendedS3DestinationUpdate_s3BackupUpdate - The Amazon S3 destination for backup.

$sel:encryptionConfiguration:ExtendedS3DestinationUpdate', extendedS3DestinationUpdate_encryptionConfiguration - The encryption configuration. If no value is specified, the default is no encryption.

$sel:dynamicPartitioningConfiguration:ExtendedS3DestinationUpdate', extendedS3DestinationUpdate_dynamicPartitioningConfiguration - The configuration of the dynamic partitioning mechanism that creates smaller data sets from the streaming data by partitioning it based on partition keys. Currently, dynamic partitioning is only supported for Amazon S3 destinations. For more information, see https://docs.aws.amazon.com/firehose/latest/dev/dynamic-partitioning.html

$sel:compressionFormat:ExtendedS3DestinationUpdate', extendedS3DestinationUpdate_compressionFormat - The compression format. If no value is specified, the default is UNCOMPRESSED.

$sel:bufferingHints:ExtendedS3DestinationUpdate', extendedS3DestinationUpdate_bufferingHints - The buffering option.

$sel:dataFormatConversionConfiguration:ExtendedS3DestinationUpdate', extendedS3DestinationUpdate_dataFormatConversionConfiguration - The serializer, deserializer, and schema for converting data from the JSON format to the Parquet or ORC format before writing it to Amazon S3.

$sel:bucketARN:ExtendedS3DestinationUpdate', extendedS3DestinationUpdate_bucketARN - The ARN of the S3 bucket. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

$sel:processingConfiguration:ExtendedS3DestinationUpdate', extendedS3DestinationUpdate_processingConfiguration - The data processing configuration.

$sel:roleARN:ExtendedS3DestinationUpdate', extendedS3DestinationUpdate_roleARN - The Amazon Resource Name (ARN) of the AWS credentials. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

extendedS3DestinationUpdate_s3BackupMode :: Lens' ExtendedS3DestinationUpdate (Maybe S3BackupMode)

You can update a delivery stream to enable Amazon S3 backup if it is disabled. If backup is enabled, you can't update the delivery stream to disable it.

extendedS3DestinationUpdate_prefix :: Lens' ExtendedS3DestinationUpdate (Maybe Text)

The "YYYY/MM/DD/HH" time format prefix is automatically used for delivered Amazon S3 files. You can also specify a custom prefix, as described in Custom Prefixes for Amazon S3 Objects.

extendedS3DestinationUpdate_cloudWatchLoggingOptions :: Lens' ExtendedS3DestinationUpdate (Maybe CloudWatchLoggingOptions)

The Amazon CloudWatch logging options for your delivery stream.

extendedS3DestinationUpdate_errorOutputPrefix :: Lens' ExtendedS3DestinationUpdate (Maybe Text)

A prefix that Kinesis Data Firehose evaluates and adds to failed records before writing them to S3. This prefix appears immediately following the bucket name. For information about how to specify this prefix, see Custom Prefixes for Amazon S3 Objects.

extendedS3DestinationUpdate_s3BackupUpdate :: Lens' ExtendedS3DestinationUpdate (Maybe S3DestinationUpdate)

The Amazon S3 destination for backup.

extendedS3DestinationUpdate_encryptionConfiguration :: Lens' ExtendedS3DestinationUpdate (Maybe EncryptionConfiguration)

The encryption configuration. If no value is specified, the default is no encryption.

extendedS3DestinationUpdate_dynamicPartitioningConfiguration :: Lens' ExtendedS3DestinationUpdate (Maybe DynamicPartitioningConfiguration)

The configuration of the dynamic partitioning mechanism that creates smaller data sets from the streaming data by partitioning it based on partition keys. Currently, dynamic partitioning is only supported for Amazon S3 destinations. For more information, see https://docs.aws.amazon.com/firehose/latest/dev/dynamic-partitioning.html

extendedS3DestinationUpdate_compressionFormat :: Lens' ExtendedS3DestinationUpdate (Maybe CompressionFormat)

The compression format. If no value is specified, the default is UNCOMPRESSED.

extendedS3DestinationUpdate_bufferingHints :: Lens' ExtendedS3DestinationUpdate (Maybe BufferingHints)

The buffering option.

extendedS3DestinationUpdate_dataFormatConversionConfiguration :: Lens' ExtendedS3DestinationUpdate (Maybe DataFormatConversionConfiguration)

The serializer, deserializer, and schema for converting data from the JSON format to the Parquet or ORC format before writing it to Amazon S3.

extendedS3DestinationUpdate_bucketARN :: Lens' ExtendedS3DestinationUpdate (Maybe Text)

The ARN of the S3 bucket. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

extendedS3DestinationUpdate_processingConfiguration :: Lens' ExtendedS3DestinationUpdate (Maybe ProcessingConfiguration)

The data processing configuration.

extendedS3DestinationUpdate_roleARN :: Lens' ExtendedS3DestinationUpdate (Maybe Text)

The Amazon Resource Name (ARN) of the AWS credentials. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces.