Network.AWS.Firehose.Types.HttpEndpointBufferingHints

Description

Documentation

data HttpEndpointBufferingHints

Describes the buffering options that can be applied before data is delivered to the HTTP endpoint destination. Kinesis Data Firehose treats these options as hints, and it might choose to use more optimal values. The SizeInMBs and IntervalInSeconds parameters are optional. However, if specify a value for one of them, you must also provide a value for the other.

See: newHttpEndpointBufferingHints smart constructor.

Constructors

HttpEndpointBufferingHints'

Fields

sizeInMBs :: Maybe Natural

Buffer incoming data to the specified size, in MBs, before delivering it to the destination. The default value is 5.

We recommend setting this parameter to a value greater than the amount of data you typically ingest into the delivery stream in 10 seconds. For example, if you typically ingest data at 1 MB/sec, the value should be 10 MB or higher.

intervalInSeconds :: Maybe Natural

Buffer incoming data for the specified period of time, in seconds, before delivering it to the destination. The default value is 300 (5 minutes).

newHttpEndpointBufferingHints :: HttpEndpointBufferingHints

Create a value of HttpEndpointBufferingHints with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sizeInMBs:HttpEndpointBufferingHints', httpEndpointBufferingHints_sizeInMBs - Buffer incoming data to the specified size, in MBs, before delivering it to the destination. The default value is 5.

We recommend setting this parameter to a value greater than the amount of data you typically ingest into the delivery stream in 10 seconds. For example, if you typically ingest data at 1 MB/sec, the value should be 10 MB or higher.

$sel:intervalInSeconds:HttpEndpointBufferingHints', httpEndpointBufferingHints_intervalInSeconds - Buffer incoming data for the specified period of time, in seconds, before delivering it to the destination. The default value is 300 (5 minutes).

httpEndpointBufferingHints_sizeInMBs :: Lens' HttpEndpointBufferingHints (Maybe Natural)

Buffer incoming data to the specified size, in MBs, before delivering it to the destination. The default value is 5.

We recommend setting this parameter to a value greater than the amount of data you typically ingest into the delivery stream in 10 seconds. For example, if you typically ingest data at 1 MB/sec, the value should be 10 MB or higher.

httpEndpointBufferingHints_intervalInSeconds :: Lens' HttpEndpointBufferingHints (Maybe Natural)

Buffer incoming data for the specified period of time, in seconds, before delivering it to the destination. The default value is 300 (5 minutes).