Network.AWS.Firehose.Types.HttpEndpointConfiguration

Description

Documentation

data HttpEndpointConfiguration

Describes the configuration of the HTTP endpoint to which Kinesis Firehose delivers data.

See: newHttpEndpointConfiguration smart constructor.

Constructors

HttpEndpointConfiguration'

Fields

accessKey :: Maybe (Sensitive Text)

The access key required for Kinesis Firehose to authenticate with the HTTP endpoint selected as the destination.

name :: Maybe Text

The name of the HTTP endpoint selected as the destination.

url :: Sensitive Text

The URL of the HTTP endpoint selected as the destination.

If you choose an HTTP endpoint as your destination, review and follow the instructions in the Appendix - HTTP Endpoint Delivery Request and Response Specifications.

newHttpEndpointConfiguration

Arguments

:: Text
$sel:url:HttpEndpointConfiguration'

-> HttpEndpointConfiguration

Create a value of HttpEndpointConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accessKey:HttpEndpointConfiguration', httpEndpointConfiguration_accessKey - The access key required for Kinesis Firehose to authenticate with the HTTP endpoint selected as the destination.

$sel:name:HttpEndpointConfiguration', httpEndpointConfiguration_name - The name of the HTTP endpoint selected as the destination.

$sel:url:HttpEndpointConfiguration', httpEndpointConfiguration_url - The URL of the HTTP endpoint selected as the destination.

If you choose an HTTP endpoint as your destination, review and follow the instructions in the Appendix - HTTP Endpoint Delivery Request and Response Specifications.

httpEndpointConfiguration_accessKey :: Lens' HttpEndpointConfiguration (Maybe Text)

The access key required for Kinesis Firehose to authenticate with the HTTP endpoint selected as the destination.

httpEndpointConfiguration_name :: Lens' HttpEndpointConfiguration (Maybe Text)

The name of the HTTP endpoint selected as the destination.

httpEndpointConfiguration_url :: Lens' HttpEndpointConfiguration Text

The URL of the HTTP endpoint selected as the destination.

If you choose an HTTP endpoint as your destination, review and follow the instructions in the Appendix - HTTP Endpoint Delivery Request and Response Specifications.