Network.AWS.Firehose.Types.HttpEndpointDestinationConfiguration

Description

Documentation

data HttpEndpointDestinationConfiguration

Describes the configuration of the HTTP endpoint destination.

See: newHttpEndpointDestinationConfiguration smart constructor.

Constructors

HttpEndpointDestinationConfiguration'

Fields

s3BackupMode :: Maybe HttpEndpointS3BackupMode

Describes the S3 bucket backup options for the data that Kinesis Data Firehose delivers to the HTTP endpoint destination. You can back up all documents (AllData) or only the documents that Kinesis Data Firehose could not deliver to the specified HTTP endpoint destination (FailedDataOnly).

cloudWatchLoggingOptions :: Maybe CloudWatchLoggingOptions

bufferingHints :: Maybe HttpEndpointBufferingHints

The buffering options that can be used before data is delivered to the specified destination. Kinesis Data Firehose treats these options as hints, and it might choose to use more optimal values. The SizeInMBs and IntervalInSeconds parameters are optional. However, if you specify a value for one of them, you must also provide a value for the other.

retryOptions :: Maybe HttpEndpointRetryOptions

Describes the retry behavior in case Kinesis Data Firehose is unable to deliver data to the specified HTTP endpoint destination, or if it doesn't receive a valid acknowledgment of receipt from the specified HTTP endpoint destination.

processingConfiguration :: Maybe ProcessingConfiguration

requestConfiguration :: Maybe HttpEndpointRequestConfiguration

The configuration of the requeste sent to the HTTP endpoint specified as the destination.

roleARN :: Maybe Text

Kinesis Data Firehose uses this IAM role for all the permissions that the delivery stream needs.

endpointConfiguration :: HttpEndpointConfiguration

The configuration of the HTTP endpoint selected as the destination.

s3Configuration :: S3DestinationConfiguration

newHttpEndpointDestinationConfiguration

Arguments

:: HttpEndpointConfiguration
$sel:endpointConfiguration:HttpEndpointDestinationConfiguration'

-> S3DestinationConfiguration
$sel:s3Configuration:HttpEndpointDestinationConfiguration'

-> HttpEndpointDestinationConfiguration

Create a value of HttpEndpointDestinationConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3BackupMode:HttpEndpointDestinationConfiguration', httpEndpointDestinationConfiguration_s3BackupMode - Describes the S3 bucket backup options for the data that Kinesis Data Firehose delivers to the HTTP endpoint destination. You can back up all documents (AllData) or only the documents that Kinesis Data Firehose could not deliver to the specified HTTP endpoint destination (FailedDataOnly).

$sel:cloudWatchLoggingOptions:HttpEndpointDestinationConfiguration', httpEndpointDestinationConfiguration_cloudWatchLoggingOptions - Undocumented member.

$sel:bufferingHints:HttpEndpointDestinationConfiguration', httpEndpointDestinationConfiguration_bufferingHints - The buffering options that can be used before data is delivered to the specified destination. Kinesis Data Firehose treats these options as hints, and it might choose to use more optimal values. The SizeInMBs and IntervalInSeconds parameters are optional. However, if you specify a value for one of them, you must also provide a value for the other.

$sel:retryOptions:HttpEndpointDestinationConfiguration', httpEndpointDestinationConfiguration_retryOptions - Describes the retry behavior in case Kinesis Data Firehose is unable to deliver data to the specified HTTP endpoint destination, or if it doesn't receive a valid acknowledgment of receipt from the specified HTTP endpoint destination.

$sel:processingConfiguration:HttpEndpointDestinationConfiguration', httpEndpointDestinationConfiguration_processingConfiguration - Undocumented member.

$sel:requestConfiguration:HttpEndpointDestinationConfiguration', httpEndpointDestinationConfiguration_requestConfiguration - The configuration of the requeste sent to the HTTP endpoint specified as the destination.

$sel:roleARN:HttpEndpointDestinationConfiguration', httpEndpointDestinationConfiguration_roleARN - Kinesis Data Firehose uses this IAM role for all the permissions that the delivery stream needs.

$sel:endpointConfiguration:HttpEndpointDestinationConfiguration', httpEndpointDestinationConfiguration_endpointConfiguration - The configuration of the HTTP endpoint selected as the destination.

$sel:s3Configuration:HttpEndpointDestinationConfiguration', httpEndpointDestinationConfiguration_s3Configuration - Undocumented member.

httpEndpointDestinationConfiguration_s3BackupMode :: Lens' HttpEndpointDestinationConfiguration (Maybe HttpEndpointS3BackupMode)

Describes the S3 bucket backup options for the data that Kinesis Data Firehose delivers to the HTTP endpoint destination. You can back up all documents (AllData) or only the documents that Kinesis Data Firehose could not deliver to the specified HTTP endpoint destination (FailedDataOnly).

httpEndpointDestinationConfiguration_cloudWatchLoggingOptions :: Lens' HttpEndpointDestinationConfiguration (Maybe CloudWatchLoggingOptions)

Undocumented member.

httpEndpointDestinationConfiguration_bufferingHints :: Lens' HttpEndpointDestinationConfiguration (Maybe HttpEndpointBufferingHints)

The buffering options that can be used before data is delivered to the specified destination. Kinesis Data Firehose treats these options as hints, and it might choose to use more optimal values. The SizeInMBs and IntervalInSeconds parameters are optional. However, if you specify a value for one of them, you must also provide a value for the other.

httpEndpointDestinationConfiguration_retryOptions :: Lens' HttpEndpointDestinationConfiguration (Maybe HttpEndpointRetryOptions)

Describes the retry behavior in case Kinesis Data Firehose is unable to deliver data to the specified HTTP endpoint destination, or if it doesn't receive a valid acknowledgment of receipt from the specified HTTP endpoint destination.

httpEndpointDestinationConfiguration_processingConfiguration :: Lens' HttpEndpointDestinationConfiguration (Maybe ProcessingConfiguration)

Undocumented member.

httpEndpointDestinationConfiguration_requestConfiguration :: Lens' HttpEndpointDestinationConfiguration (Maybe HttpEndpointRequestConfiguration)

The configuration of the requeste sent to the HTTP endpoint specified as the destination.

httpEndpointDestinationConfiguration_roleARN :: Lens' HttpEndpointDestinationConfiguration (Maybe Text)

Kinesis Data Firehose uses this IAM role for all the permissions that the delivery stream needs.

httpEndpointDestinationConfiguration_endpointConfiguration :: Lens' HttpEndpointDestinationConfiguration HttpEndpointConfiguration

The configuration of the HTTP endpoint selected as the destination.

httpEndpointDestinationConfiguration_s3Configuration :: Lens' HttpEndpointDestinationConfiguration S3DestinationConfiguration

Undocumented member.