Network.AWS.Firehose.Types.HttpEndpointDestinationDescription

Description

Documentation

data HttpEndpointDestinationDescription

Describes the HTTP endpoint destination.

See: newHttpEndpointDestinationDescription smart constructor.

Constructors

HttpEndpointDestinationDescription'

Fields

roleARN :: Maybe Text

Kinesis Data Firehose uses this IAM role for all the permissions that the delivery stream needs.

endpointConfiguration :: Maybe HttpEndpointDescription

The configuration of the specified HTTP endpoint destination.

processingConfiguration :: Maybe ProcessingConfiguration

cloudWatchLoggingOptions :: Maybe CloudWatchLoggingOptions

requestConfiguration :: Maybe HttpEndpointRequestConfiguration

The configuration of request sent to the HTTP endpoint specified as the destination.

bufferingHints :: Maybe HttpEndpointBufferingHints

Describes buffering options that can be applied to the data before it is delivered to the HTTPS endpoint destination. Kinesis Data Firehose teats these options as hints, and it might choose to use more optimal values. The SizeInMBs and IntervalInSeconds parameters are optional. However, if specify a value for one of them, you must also provide a value for the other.

retryOptions :: Maybe HttpEndpointRetryOptions

Describes the retry behavior in case Kinesis Data Firehose is unable to deliver data to the specified HTTP endpoint destination, or if it doesn't receive a valid acknowledgment of receipt from the specified HTTP endpoint destination.

s3BackupMode :: Maybe HttpEndpointS3BackupMode

Describes the S3 bucket backup options for the data that Kinesis Firehose delivers to the HTTP endpoint destination. You can back up all documents (AllData) or only the documents that Kinesis Data Firehose could not deliver to the specified HTTP endpoint destination (FailedDataOnly).

s3DestinationDescription :: Maybe S3DestinationDescription

newHttpEndpointDestinationDescription :: HttpEndpointDestinationDescription

Create a value of HttpEndpointDestinationDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:roleARN:HttpEndpointDestinationDescription', httpEndpointDestinationDescription_roleARN - Kinesis Data Firehose uses this IAM role for all the permissions that the delivery stream needs.

$sel:endpointConfiguration:HttpEndpointDestinationDescription', httpEndpointDestinationDescription_endpointConfiguration - The configuration of the specified HTTP endpoint destination.

$sel:processingConfiguration:HttpEndpointDestinationDescription', httpEndpointDestinationDescription_processingConfiguration - Undocumented member.

$sel:cloudWatchLoggingOptions:HttpEndpointDestinationDescription', httpEndpointDestinationDescription_cloudWatchLoggingOptions - Undocumented member.

$sel:requestConfiguration:HttpEndpointDestinationDescription', httpEndpointDestinationDescription_requestConfiguration - The configuration of request sent to the HTTP endpoint specified as the destination.

$sel:bufferingHints:HttpEndpointDestinationDescription', httpEndpointDestinationDescription_bufferingHints - Describes buffering options that can be applied to the data before it is delivered to the HTTPS endpoint destination. Kinesis Data Firehose teats these options as hints, and it might choose to use more optimal values. The SizeInMBs and IntervalInSeconds parameters are optional. However, if specify a value for one of them, you must also provide a value for the other.

$sel:retryOptions:HttpEndpointDestinationDescription', httpEndpointDestinationDescription_retryOptions - Describes the retry behavior in case Kinesis Data Firehose is unable to deliver data to the specified HTTP endpoint destination, or if it doesn't receive a valid acknowledgment of receipt from the specified HTTP endpoint destination.

$sel:s3BackupMode:HttpEndpointDestinationDescription', httpEndpointDestinationDescription_s3BackupMode - Describes the S3 bucket backup options for the data that Kinesis Firehose delivers to the HTTP endpoint destination. You can back up all documents (AllData) or only the documents that Kinesis Data Firehose could not deliver to the specified HTTP endpoint destination (FailedDataOnly).

$sel:s3DestinationDescription:HttpEndpointDestinationDescription', httpEndpointDestinationDescription_s3DestinationDescription - Undocumented member.

httpEndpointDestinationDescription_roleARN :: Lens' HttpEndpointDestinationDescription (Maybe Text)

Kinesis Data Firehose uses this IAM role for all the permissions that the delivery stream needs.

httpEndpointDestinationDescription_endpointConfiguration :: Lens' HttpEndpointDestinationDescription (Maybe HttpEndpointDescription)

The configuration of the specified HTTP endpoint destination.

httpEndpointDestinationDescription_processingConfiguration :: Lens' HttpEndpointDestinationDescription (Maybe ProcessingConfiguration)

Undocumented member.

httpEndpointDestinationDescription_cloudWatchLoggingOptions :: Lens' HttpEndpointDestinationDescription (Maybe CloudWatchLoggingOptions)

Undocumented member.

httpEndpointDestinationDescription_requestConfiguration :: Lens' HttpEndpointDestinationDescription (Maybe HttpEndpointRequestConfiguration)

The configuration of request sent to the HTTP endpoint specified as the destination.

httpEndpointDestinationDescription_bufferingHints :: Lens' HttpEndpointDestinationDescription (Maybe HttpEndpointBufferingHints)

Describes buffering options that can be applied to the data before it is delivered to the HTTPS endpoint destination. Kinesis Data Firehose teats these options as hints, and it might choose to use more optimal values. The SizeInMBs and IntervalInSeconds parameters are optional. However, if specify a value for one of them, you must also provide a value for the other.

httpEndpointDestinationDescription_retryOptions :: Lens' HttpEndpointDestinationDescription (Maybe HttpEndpointRetryOptions)

Describes the retry behavior in case Kinesis Data Firehose is unable to deliver data to the specified HTTP endpoint destination, or if it doesn't receive a valid acknowledgment of receipt from the specified HTTP endpoint destination.

httpEndpointDestinationDescription_s3BackupMode :: Lens' HttpEndpointDestinationDescription (Maybe HttpEndpointS3BackupMode)

Describes the S3 bucket backup options for the data that Kinesis Firehose delivers to the HTTP endpoint destination. You can back up all documents (AllData) or only the documents that Kinesis Data Firehose could not deliver to the specified HTTP endpoint destination (FailedDataOnly).

httpEndpointDestinationDescription_s3DestinationDescription :: Lens' HttpEndpointDestinationDescription (Maybe S3DestinationDescription)

Undocumented member.