Network.AWS.Firehose.Types.HttpEndpointRequestConfiguration

Description

Documentation

data HttpEndpointRequestConfiguration

The configuration of the HTTP endpoint request.

See: newHttpEndpointRequestConfiguration smart constructor.

Constructors

HttpEndpointRequestConfiguration'

Fields

commonAttributes :: Maybe [HttpEndpointCommonAttribute]

Describes the metadata sent to the HTTP endpoint destination.

contentEncoding :: Maybe ContentEncoding

Kinesis Data Firehose uses the content encoding to compress the body of a request before sending the request to the destination. For more information, see Content-Encoding in MDN Web Docs, the official Mozilla documentation.

newHttpEndpointRequestConfiguration :: HttpEndpointRequestConfiguration

Create a value of HttpEndpointRequestConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:commonAttributes:HttpEndpointRequestConfiguration', httpEndpointRequestConfiguration_commonAttributes - Describes the metadata sent to the HTTP endpoint destination.

$sel:contentEncoding:HttpEndpointRequestConfiguration', httpEndpointRequestConfiguration_contentEncoding - Kinesis Data Firehose uses the content encoding to compress the body of a request before sending the request to the destination. For more information, see Content-Encoding in MDN Web Docs, the official Mozilla documentation.

httpEndpointRequestConfiguration_commonAttributes :: Lens' HttpEndpointRequestConfiguration (Maybe [HttpEndpointCommonAttribute])

Describes the metadata sent to the HTTP endpoint destination.

httpEndpointRequestConfiguration_contentEncoding :: Lens' HttpEndpointRequestConfiguration (Maybe ContentEncoding)

Kinesis Data Firehose uses the content encoding to compress the body of a request before sending the request to the destination. For more information, see Content-Encoding in MDN Web Docs, the official Mozilla documentation.