Network.AWS.Firehose.Types.ParquetSerDe

Description

Documentation

data ParquetSerDe

A serializer to use for converting data to the Parquet format before storing it in Amazon S3. For more information, see Apache Parquet.

See: newParquetSerDe smart constructor.

Constructors

ParquetSerDe'

Fields

writerVersion :: Maybe ParquetWriterVersion

Indicates the version of row format to output. The possible values are V1 and V2. The default is V1.

compression :: Maybe ParquetCompression

The compression code to use over data blocks. The possible values are UNCOMPRESSED, SNAPPY, and GZIP, with the default being SNAPPY. Use SNAPPY for higher decompression speed. Use GZIP if the compression ratio is more important than speed.

maxPaddingBytes :: Maybe Natural

The maximum amount of padding to apply. This is useful if you intend to copy the data from Amazon S3 to HDFS before querying. The default is 0.

enableDictionaryCompression :: Maybe Bool

Indicates whether to enable dictionary compression.

pageSizeBytes :: Maybe Natural

The Parquet page size. Column chunks are divided into pages. A page is conceptually an indivisible unit (in terms of compression and encoding). The minimum value is 64 KiB and the default is 1 MiB.

blockSizeBytes :: Maybe Natural

The Hadoop Distributed File System (HDFS) block size. This is useful if you intend to copy the data from Amazon S3 to HDFS before querying. The default is 256 MiB and the minimum is 64 MiB. Kinesis Data Firehose uses this value for padding calculations.

newParquetSerDe :: ParquetSerDe

Create a value of ParquetSerDe with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:writerVersion:ParquetSerDe', parquetSerDe_writerVersion - Indicates the version of row format to output. The possible values are V1 and V2. The default is V1.

$sel:compression:ParquetSerDe', parquetSerDe_compression - The compression code to use over data blocks. The possible values are UNCOMPRESSED, SNAPPY, and GZIP, with the default being SNAPPY. Use SNAPPY for higher decompression speed. Use GZIP if the compression ratio is more important than speed.

$sel:maxPaddingBytes:ParquetSerDe', parquetSerDe_maxPaddingBytes - The maximum amount of padding to apply. This is useful if you intend to copy the data from Amazon S3 to HDFS before querying. The default is 0.

$sel:enableDictionaryCompression:ParquetSerDe', parquetSerDe_enableDictionaryCompression - Indicates whether to enable dictionary compression.

$sel:pageSizeBytes:ParquetSerDe', parquetSerDe_pageSizeBytes - The Parquet page size. Column chunks are divided into pages. A page is conceptually an indivisible unit (in terms of compression and encoding). The minimum value is 64 KiB and the default is 1 MiB.

$sel:blockSizeBytes:ParquetSerDe', parquetSerDe_blockSizeBytes - The Hadoop Distributed File System (HDFS) block size. This is useful if you intend to copy the data from Amazon S3 to HDFS before querying. The default is 256 MiB and the minimum is 64 MiB. Kinesis Data Firehose uses this value for padding calculations.

parquetSerDe_writerVersion :: Lens' ParquetSerDe (Maybe ParquetWriterVersion)

Indicates the version of row format to output. The possible values are V1 and V2. The default is V1.

parquetSerDe_compression :: Lens' ParquetSerDe (Maybe ParquetCompression)

The compression code to use over data blocks. The possible values are UNCOMPRESSED, SNAPPY, and GZIP, with the default being SNAPPY. Use SNAPPY for higher decompression speed. Use GZIP if the compression ratio is more important than speed.

parquetSerDe_maxPaddingBytes :: Lens' ParquetSerDe (Maybe Natural)

The maximum amount of padding to apply. This is useful if you intend to copy the data from Amazon S3 to HDFS before querying. The default is 0.

parquetSerDe_enableDictionaryCompression :: Lens' ParquetSerDe (Maybe Bool)

Indicates whether to enable dictionary compression.

parquetSerDe_pageSizeBytes :: Lens' ParquetSerDe (Maybe Natural)

The Parquet page size. Column chunks are divided into pages. A page is conceptually an indivisible unit (in terms of compression and encoding). The minimum value is 64 KiB and the default is 1 MiB.

parquetSerDe_blockSizeBytes :: Lens' ParquetSerDe (Maybe Natural)

The Hadoop Distributed File System (HDFS) block size. This is useful if you intend to copy the data from Amazon S3 to HDFS before querying. The default is 256 MiB and the minimum is 64 MiB. Kinesis Data Firehose uses this value for padding calculations.