Network.AWS.Firehose.Types.RedshiftDestinationConfiguration

Description

Documentation

data RedshiftDestinationConfiguration

Describes the configuration of a destination in Amazon Redshift.

See: newRedshiftDestinationConfiguration smart constructor.

Constructors

RedshiftDestinationConfiguration'

Fields

s3BackupConfiguration :: Maybe S3DestinationConfiguration

The configuration for backup in Amazon S3.

processingConfiguration :: Maybe ProcessingConfiguration

The data processing configuration.

cloudWatchLoggingOptions :: Maybe CloudWatchLoggingOptions

The CloudWatch logging options for your delivery stream.

retryOptions :: Maybe RedshiftRetryOptions

The retry behavior in case Kinesis Data Firehose is unable to deliver documents to Amazon Redshift. Default value is 3600 (60 minutes).

s3BackupMode :: Maybe RedshiftS3BackupMode

The Amazon S3 backup mode. After you create a delivery stream, you can update it to enable Amazon S3 backup if it is disabled. If backup is enabled, you can't update the delivery stream to disable it.

roleARN :: Text

The Amazon Resource Name (ARN) of the AWS credentials. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

clusterJDBCURL :: Text

The database connection string.

copyCommand :: CopyCommand

The COPY command.

username :: Sensitive Text

The name of the user.

password :: Sensitive Text

The user password.

s3Configuration :: S3DestinationConfiguration

The configuration for the intermediate Amazon S3 location from which Amazon Redshift obtains data. Restrictions are described in the topic for CreateDeliveryStream.

The compression formats SNAPPY or ZIP cannot be specified in RedshiftDestinationConfiguration.S3Configuration because the Amazon Redshift COPY operation that reads from the S3 bucket doesn't support these compression formats.

newRedshiftDestinationConfiguration

Arguments

:: Text
$sel:roleARN:RedshiftDestinationConfiguration'

-> Text
$sel:clusterJDBCURL:RedshiftDestinationConfiguration'

-> CopyCommand
$sel:copyCommand:RedshiftDestinationConfiguration'

-> Text
$sel:username:RedshiftDestinationConfiguration'

-> Text
$sel:password:RedshiftDestinationConfiguration'

-> S3DestinationConfiguration
$sel:s3Configuration:RedshiftDestinationConfiguration'

-> RedshiftDestinationConfiguration

Create a value of RedshiftDestinationConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3BackupConfiguration:RedshiftDestinationConfiguration', redshiftDestinationConfiguration_s3BackupConfiguration - The configuration for backup in Amazon S3.

$sel:processingConfiguration:RedshiftDestinationConfiguration', redshiftDestinationConfiguration_processingConfiguration - The data processing configuration.

$sel:cloudWatchLoggingOptions:RedshiftDestinationConfiguration', redshiftDestinationConfiguration_cloudWatchLoggingOptions - The CloudWatch logging options for your delivery stream.

$sel:retryOptions:RedshiftDestinationConfiguration', redshiftDestinationConfiguration_retryOptions - The retry behavior in case Kinesis Data Firehose is unable to deliver documents to Amazon Redshift. Default value is 3600 (60 minutes).

$sel:s3BackupMode:RedshiftDestinationConfiguration', redshiftDestinationConfiguration_s3BackupMode - The Amazon S3 backup mode. After you create a delivery stream, you can update it to enable Amazon S3 backup if it is disabled. If backup is enabled, you can't update the delivery stream to disable it.

$sel:roleARN:RedshiftDestinationConfiguration', redshiftDestinationConfiguration_roleARN - The Amazon Resource Name (ARN) of the AWS credentials. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

$sel:clusterJDBCURL:RedshiftDestinationConfiguration', redshiftDestinationConfiguration_clusterJDBCURL - The database connection string.

$sel:copyCommand:RedshiftDestinationConfiguration', redshiftDestinationConfiguration_copyCommand - The COPY command.

$sel:username:RedshiftDestinationConfiguration', redshiftDestinationConfiguration_username - The name of the user.

$sel:password:RedshiftDestinationConfiguration', redshiftDestinationConfiguration_password - The user password.

$sel:s3Configuration:RedshiftDestinationConfiguration', redshiftDestinationConfiguration_s3Configuration - The configuration for the intermediate Amazon S3 location from which Amazon Redshift obtains data. Restrictions are described in the topic for CreateDeliveryStream.

The compression formats SNAPPY or ZIP cannot be specified in RedshiftDestinationConfiguration.S3Configuration because the Amazon Redshift COPY operation that reads from the S3 bucket doesn't support these compression formats.

redshiftDestinationConfiguration_s3BackupConfiguration :: Lens' RedshiftDestinationConfiguration (Maybe S3DestinationConfiguration)

The configuration for backup in Amazon S3.

redshiftDestinationConfiguration_processingConfiguration :: Lens' RedshiftDestinationConfiguration (Maybe ProcessingConfiguration)

The data processing configuration.

redshiftDestinationConfiguration_cloudWatchLoggingOptions :: Lens' RedshiftDestinationConfiguration (Maybe CloudWatchLoggingOptions)

The CloudWatch logging options for your delivery stream.

redshiftDestinationConfiguration_retryOptions :: Lens' RedshiftDestinationConfiguration (Maybe RedshiftRetryOptions)

The retry behavior in case Kinesis Data Firehose is unable to deliver documents to Amazon Redshift. Default value is 3600 (60 minutes).

redshiftDestinationConfiguration_s3BackupMode :: Lens' RedshiftDestinationConfiguration (Maybe RedshiftS3BackupMode)

The Amazon S3 backup mode. After you create a delivery stream, you can update it to enable Amazon S3 backup if it is disabled. If backup is enabled, you can't update the delivery stream to disable it.

redshiftDestinationConfiguration_roleARN :: Lens' RedshiftDestinationConfiguration Text

The Amazon Resource Name (ARN) of the AWS credentials. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

redshiftDestinationConfiguration_clusterJDBCURL :: Lens' RedshiftDestinationConfiguration Text

The database connection string.

redshiftDestinationConfiguration_copyCommand :: Lens' RedshiftDestinationConfiguration CopyCommand

The COPY command.

redshiftDestinationConfiguration_username :: Lens' RedshiftDestinationConfiguration Text

The name of the user.

redshiftDestinationConfiguration_password :: Lens' RedshiftDestinationConfiguration Text

The user password.

redshiftDestinationConfiguration_s3Configuration :: Lens' RedshiftDestinationConfiguration S3DestinationConfiguration

The configuration for the intermediate Amazon S3 location from which Amazon Redshift obtains data. Restrictions are described in the topic for CreateDeliveryStream.

The compression formats SNAPPY or ZIP cannot be specified in RedshiftDestinationConfiguration.S3Configuration because the Amazon Redshift COPY operation that reads from the S3 bucket doesn't support these compression formats.