Network.AWS.Firehose.Types.RedshiftDestinationDescription

Description

Documentation

data RedshiftDestinationDescription

Describes a destination in Amazon Redshift.

See: newRedshiftDestinationDescription smart constructor.

Constructors

RedshiftDestinationDescription'

Fields

processingConfiguration :: Maybe ProcessingConfiguration

The data processing configuration.

cloudWatchLoggingOptions :: Maybe CloudWatchLoggingOptions

The Amazon CloudWatch logging options for your delivery stream.

s3BackupDescription :: Maybe S3DestinationDescription

The configuration for backup in Amazon S3.

retryOptions :: Maybe RedshiftRetryOptions

The retry behavior in case Kinesis Data Firehose is unable to deliver documents to Amazon Redshift. Default value is 3600 (60 minutes).

s3BackupMode :: Maybe RedshiftS3BackupMode

The Amazon S3 backup mode.

roleARN :: Text

The Amazon Resource Name (ARN) of the AWS credentials. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

clusterJDBCURL :: Text

The database connection string.

copyCommand :: CopyCommand

The COPY command.

username :: Sensitive Text

The name of the user.

s3DestinationDescription :: S3DestinationDescription

The Amazon S3 destination.

newRedshiftDestinationDescription

Arguments

:: Text
$sel:roleARN:RedshiftDestinationDescription'

-> Text
$sel:clusterJDBCURL:RedshiftDestinationDescription'

-> CopyCommand
$sel:copyCommand:RedshiftDestinationDescription'

-> Text
$sel:username:RedshiftDestinationDescription'

-> S3DestinationDescription
$sel:s3DestinationDescription:RedshiftDestinationDescription'

-> RedshiftDestinationDescription

Create a value of RedshiftDestinationDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:processingConfiguration:RedshiftDestinationDescription', redshiftDestinationDescription_processingConfiguration - The data processing configuration.

$sel:cloudWatchLoggingOptions:RedshiftDestinationDescription', redshiftDestinationDescription_cloudWatchLoggingOptions - The Amazon CloudWatch logging options for your delivery stream.

$sel:s3BackupDescription:RedshiftDestinationDescription', redshiftDestinationDescription_s3BackupDescription - The configuration for backup in Amazon S3.

$sel:retryOptions:RedshiftDestinationDescription', redshiftDestinationDescription_retryOptions - The retry behavior in case Kinesis Data Firehose is unable to deliver documents to Amazon Redshift. Default value is 3600 (60 minutes).

$sel:s3BackupMode:RedshiftDestinationDescription', redshiftDestinationDescription_s3BackupMode - The Amazon S3 backup mode.

$sel:roleARN:RedshiftDestinationDescription', redshiftDestinationDescription_roleARN - The Amazon Resource Name (ARN) of the AWS credentials. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

$sel:clusterJDBCURL:RedshiftDestinationDescription', redshiftDestinationDescription_clusterJDBCURL - The database connection string.

$sel:copyCommand:RedshiftDestinationDescription', redshiftDestinationDescription_copyCommand - The COPY command.

$sel:username:RedshiftDestinationDescription', redshiftDestinationDescription_username - The name of the user.

$sel:s3DestinationDescription:RedshiftDestinationDescription', redshiftDestinationDescription_s3DestinationDescription - The Amazon S3 destination.

redshiftDestinationDescription_processingConfiguration :: Lens' RedshiftDestinationDescription (Maybe ProcessingConfiguration)

The data processing configuration.

redshiftDestinationDescription_cloudWatchLoggingOptions :: Lens' RedshiftDestinationDescription (Maybe CloudWatchLoggingOptions)

The Amazon CloudWatch logging options for your delivery stream.

redshiftDestinationDescription_s3BackupDescription :: Lens' RedshiftDestinationDescription (Maybe S3DestinationDescription)

The configuration for backup in Amazon S3.

redshiftDestinationDescription_retryOptions :: Lens' RedshiftDestinationDescription (Maybe RedshiftRetryOptions)

The retry behavior in case Kinesis Data Firehose is unable to deliver documents to Amazon Redshift. Default value is 3600 (60 minutes).

redshiftDestinationDescription_s3BackupMode :: Lens' RedshiftDestinationDescription (Maybe RedshiftS3BackupMode)

The Amazon S3 backup mode.

redshiftDestinationDescription_roleARN :: Lens' RedshiftDestinationDescription Text

The Amazon Resource Name (ARN) of the AWS credentials. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

redshiftDestinationDescription_clusterJDBCURL :: Lens' RedshiftDestinationDescription Text

The database connection string.

redshiftDestinationDescription_copyCommand :: Lens' RedshiftDestinationDescription CopyCommand

The COPY command.

redshiftDestinationDescription_username :: Lens' RedshiftDestinationDescription Text

The name of the user.

redshiftDestinationDescription_s3DestinationDescription :: Lens' RedshiftDestinationDescription S3DestinationDescription

The Amazon S3 destination.