Network.AWS.Firehose.Types.RedshiftDestinationUpdate

Description

Documentation

data RedshiftDestinationUpdate

Describes an update for a destination in Amazon Redshift.

See: newRedshiftDestinationUpdate smart constructor.

Constructors

RedshiftDestinationUpdate'

Fields

s3BackupMode :: Maybe RedshiftS3BackupMode

You can update a delivery stream to enable Amazon S3 backup if it is disabled. If backup is enabled, you can't update the delivery stream to disable it.

cloudWatchLoggingOptions :: Maybe CloudWatchLoggingOptions

The Amazon CloudWatch logging options for your delivery stream.

username :: Maybe (Sensitive Text)

The name of the user.

s3Update :: Maybe S3DestinationUpdate

The Amazon S3 destination.

The compression formats SNAPPY or ZIP cannot be specified in RedshiftDestinationUpdate.S3Update because the Amazon Redshift COPY operation that reads from the S3 bucket doesn't support these compression formats.

password :: Maybe (Sensitive Text)

The user password.

s3BackupUpdate :: Maybe S3DestinationUpdate

The Amazon S3 destination for backup.

copyCommand :: Maybe CopyCommand

The COPY command.

retryOptions :: Maybe RedshiftRetryOptions

The retry behavior in case Kinesis Data Firehose is unable to deliver documents to Amazon Redshift. Default value is 3600 (60 minutes).

processingConfiguration :: Maybe ProcessingConfiguration

The data processing configuration.

clusterJDBCURL :: Maybe Text

The database connection string.

roleARN :: Maybe Text

The Amazon Resource Name (ARN) of the AWS credentials. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

newRedshiftDestinationUpdate :: RedshiftDestinationUpdate

Create a value of RedshiftDestinationUpdate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3BackupMode:RedshiftDestinationUpdate', redshiftDestinationUpdate_s3BackupMode - You can update a delivery stream to enable Amazon S3 backup if it is disabled. If backup is enabled, you can't update the delivery stream to disable it.

$sel:cloudWatchLoggingOptions:RedshiftDestinationUpdate', redshiftDestinationUpdate_cloudWatchLoggingOptions - The Amazon CloudWatch logging options for your delivery stream.

$sel:username:RedshiftDestinationUpdate', redshiftDestinationUpdate_username - The name of the user.

$sel:s3Update:RedshiftDestinationUpdate', redshiftDestinationUpdate_s3Update - The Amazon S3 destination.

The compression formats SNAPPY or ZIP cannot be specified in RedshiftDestinationUpdate.S3Update because the Amazon Redshift COPY operation that reads from the S3 bucket doesn't support these compression formats.

$sel:password:RedshiftDestinationUpdate', redshiftDestinationUpdate_password - The user password.

$sel:s3BackupUpdate:RedshiftDestinationUpdate', redshiftDestinationUpdate_s3BackupUpdate - The Amazon S3 destination for backup.

$sel:copyCommand:RedshiftDestinationUpdate', redshiftDestinationUpdate_copyCommand - The COPY command.

$sel:retryOptions:RedshiftDestinationUpdate', redshiftDestinationUpdate_retryOptions - The retry behavior in case Kinesis Data Firehose is unable to deliver documents to Amazon Redshift. Default value is 3600 (60 minutes).

$sel:processingConfiguration:RedshiftDestinationUpdate', redshiftDestinationUpdate_processingConfiguration - The data processing configuration.

$sel:clusterJDBCURL:RedshiftDestinationUpdate', redshiftDestinationUpdate_clusterJDBCURL - The database connection string.

$sel:roleARN:RedshiftDestinationUpdate', redshiftDestinationUpdate_roleARN - The Amazon Resource Name (ARN) of the AWS credentials. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

redshiftDestinationUpdate_s3BackupMode :: Lens' RedshiftDestinationUpdate (Maybe RedshiftS3BackupMode)

You can update a delivery stream to enable Amazon S3 backup if it is disabled. If backup is enabled, you can't update the delivery stream to disable it.

redshiftDestinationUpdate_cloudWatchLoggingOptions :: Lens' RedshiftDestinationUpdate (Maybe CloudWatchLoggingOptions)

The Amazon CloudWatch logging options for your delivery stream.

redshiftDestinationUpdate_username :: Lens' RedshiftDestinationUpdate (Maybe Text)

The name of the user.

redshiftDestinationUpdate_s3Update :: Lens' RedshiftDestinationUpdate (Maybe S3DestinationUpdate)

The Amazon S3 destination.

The compression formats SNAPPY or ZIP cannot be specified in RedshiftDestinationUpdate.S3Update because the Amazon Redshift COPY operation that reads from the S3 bucket doesn't support these compression formats.

redshiftDestinationUpdate_password :: Lens' RedshiftDestinationUpdate (Maybe Text)

The user password.

redshiftDestinationUpdate_s3BackupUpdate :: Lens' RedshiftDestinationUpdate (Maybe S3DestinationUpdate)

The Amazon S3 destination for backup.

redshiftDestinationUpdate_copyCommand :: Lens' RedshiftDestinationUpdate (Maybe CopyCommand)

The COPY command.

redshiftDestinationUpdate_retryOptions :: Lens' RedshiftDestinationUpdate (Maybe RedshiftRetryOptions)

The retry behavior in case Kinesis Data Firehose is unable to deliver documents to Amazon Redshift. Default value is 3600 (60 minutes).

redshiftDestinationUpdate_processingConfiguration :: Lens' RedshiftDestinationUpdate (Maybe ProcessingConfiguration)

The data processing configuration.

redshiftDestinationUpdate_clusterJDBCURL :: Lens' RedshiftDestinationUpdate (Maybe Text)

The database connection string.

redshiftDestinationUpdate_roleARN :: Lens' RedshiftDestinationUpdate (Maybe Text)

The Amazon Resource Name (ARN) of the AWS credentials. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces.