Network.AWS.Firehose.Types.S3DestinationConfiguration

Description

Documentation

data S3DestinationConfiguration

Describes the configuration of a destination in Amazon S3.

See: newS3DestinationConfiguration smart constructor.

Constructors

S3DestinationConfiguration'

Fields

prefix :: Maybe Text

The "YYYY/MM/DD/HH" time format prefix is automatically used for delivered Amazon S3 files. You can also specify a custom prefix, as described in Custom Prefixes for Amazon S3 Objects.

cloudWatchLoggingOptions :: Maybe CloudWatchLoggingOptions

The CloudWatch logging options for your delivery stream.

errorOutputPrefix :: Maybe Text

A prefix that Kinesis Data Firehose evaluates and adds to failed records before writing them to S3. This prefix appears immediately following the bucket name. For information about how to specify this prefix, see Custom Prefixes for Amazon S3 Objects.

encryptionConfiguration :: Maybe EncryptionConfiguration

The encryption configuration. If no value is specified, the default is no encryption.

compressionFormat :: Maybe CompressionFormat

The compression format. If no value is specified, the default is UNCOMPRESSED.

The compression formats SNAPPY or ZIP cannot be specified for Amazon Redshift destinations because they are not supported by the Amazon Redshift COPY operation that reads from the S3 bucket.

bufferingHints :: Maybe BufferingHints

The buffering option. If no value is specified, BufferingHints object default values are used.

roleARN :: Text

The Amazon Resource Name (ARN) of the AWS credentials. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

bucketARN :: Text

The ARN of the S3 bucket. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

newS3DestinationConfiguration

Arguments

:: Text
$sel:roleARN:S3DestinationConfiguration'

-> Text
$sel:bucketARN:S3DestinationConfiguration'

-> S3DestinationConfiguration

Create a value of S3DestinationConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:prefix:S3DestinationConfiguration', s3DestinationConfiguration_prefix - The "YYYY/MM/DD/HH" time format prefix is automatically used for delivered Amazon S3 files. You can also specify a custom prefix, as described in Custom Prefixes for Amazon S3 Objects.

$sel:cloudWatchLoggingOptions:S3DestinationConfiguration', s3DestinationConfiguration_cloudWatchLoggingOptions - The CloudWatch logging options for your delivery stream.

$sel:errorOutputPrefix:S3DestinationConfiguration', s3DestinationConfiguration_errorOutputPrefix - A prefix that Kinesis Data Firehose evaluates and adds to failed records before writing them to S3. This prefix appears immediately following the bucket name. For information about how to specify this prefix, see Custom Prefixes for Amazon S3 Objects.

$sel:encryptionConfiguration:S3DestinationConfiguration', s3DestinationConfiguration_encryptionConfiguration - The encryption configuration. If no value is specified, the default is no encryption.

$sel:compressionFormat:S3DestinationConfiguration', s3DestinationConfiguration_compressionFormat - The compression format. If no value is specified, the default is UNCOMPRESSED.

The compression formats SNAPPY or ZIP cannot be specified for Amazon Redshift destinations because they are not supported by the Amazon Redshift COPY operation that reads from the S3 bucket.

$sel:bufferingHints:S3DestinationConfiguration', s3DestinationConfiguration_bufferingHints - The buffering option. If no value is specified, BufferingHints object default values are used.

$sel:roleARN:S3DestinationConfiguration', s3DestinationConfiguration_roleARN - The Amazon Resource Name (ARN) of the AWS credentials. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

$sel:bucketARN:S3DestinationConfiguration', s3DestinationConfiguration_bucketARN - The ARN of the S3 bucket. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

s3DestinationConfiguration_prefix :: Lens' S3DestinationConfiguration (Maybe Text)

The "YYYY/MM/DD/HH" time format prefix is automatically used for delivered Amazon S3 files. You can also specify a custom prefix, as described in Custom Prefixes for Amazon S3 Objects.

s3DestinationConfiguration_cloudWatchLoggingOptions :: Lens' S3DestinationConfiguration (Maybe CloudWatchLoggingOptions)

The CloudWatch logging options for your delivery stream.

s3DestinationConfiguration_errorOutputPrefix :: Lens' S3DestinationConfiguration (Maybe Text)

A prefix that Kinesis Data Firehose evaluates and adds to failed records before writing them to S3. This prefix appears immediately following the bucket name. For information about how to specify this prefix, see Custom Prefixes for Amazon S3 Objects.

s3DestinationConfiguration_encryptionConfiguration :: Lens' S3DestinationConfiguration (Maybe EncryptionConfiguration)

The encryption configuration. If no value is specified, the default is no encryption.

s3DestinationConfiguration_compressionFormat :: Lens' S3DestinationConfiguration (Maybe CompressionFormat)

The compression format. If no value is specified, the default is UNCOMPRESSED.

The compression formats SNAPPY or ZIP cannot be specified for Amazon Redshift destinations because they are not supported by the Amazon Redshift COPY operation that reads from the S3 bucket.

s3DestinationConfiguration_bufferingHints :: Lens' S3DestinationConfiguration (Maybe BufferingHints)

The buffering option. If no value is specified, BufferingHints object default values are used.

s3DestinationConfiguration_roleARN :: Lens' S3DestinationConfiguration Text

The Amazon Resource Name (ARN) of the AWS credentials. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

s3DestinationConfiguration_bucketARN :: Lens' S3DestinationConfiguration Text

The ARN of the S3 bucket. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces.