Network.AWS.Firehose.Types.SchemaConfiguration

Description

Documentation

data SchemaConfiguration

Specifies the schema to which you want Kinesis Data Firehose to configure your data before it writes it to Amazon S3. This parameter is required if Enabled is set to true.

See: newSchemaConfiguration smart constructor.

Constructors

SchemaConfiguration'

Fields

versionId :: Maybe Text

Specifies the table version for the output data schema. If you don't specify this version ID, or if you set it to LATEST, Kinesis Data Firehose uses the most recent version. This means that any updates to the table are automatically picked up.

catalogId :: Maybe Text

The ID of the AWS Glue Data Catalog. If you don't supply this, the AWS account ID is used by default.

databaseName :: Maybe Text

Specifies the name of the AWS Glue database that contains the schema for the output data.

If the SchemaConfiguration request parameter is used as part of invoking the CreateDeliveryStream API, then the DatabaseName property is required and its value must be specified.

region :: Maybe Text

If you don't specify an AWS Region, the default is the current Region.

tableName :: Maybe Text

Specifies the AWS Glue table that contains the column information that constitutes your data schema.

If the SchemaConfiguration request parameter is used as part of invoking the CreateDeliveryStream API, then the TableName property is required and its value must be specified.

roleARN :: Maybe Text

The role that Kinesis Data Firehose can use to access AWS Glue. This role must be in the same account you use for Kinesis Data Firehose. Cross-account roles aren't allowed.

If the SchemaConfiguration request parameter is used as part of invoking the CreateDeliveryStream API, then the RoleARN property is required and its value must be specified.

newSchemaConfiguration :: SchemaConfiguration

Create a value of SchemaConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:versionId:SchemaConfiguration', schemaConfiguration_versionId - Specifies the table version for the output data schema. If you don't specify this version ID, or if you set it to LATEST, Kinesis Data Firehose uses the most recent version. This means that any updates to the table are automatically picked up.

$sel:catalogId:SchemaConfiguration', schemaConfiguration_catalogId - The ID of the AWS Glue Data Catalog. If you don't supply this, the AWS account ID is used by default.

$sel:databaseName:SchemaConfiguration', schemaConfiguration_databaseName - Specifies the name of the AWS Glue database that contains the schema for the output data.

If the SchemaConfiguration request parameter is used as part of invoking the CreateDeliveryStream API, then the DatabaseName property is required and its value must be specified.

$sel:region:SchemaConfiguration', schemaConfiguration_region - If you don't specify an AWS Region, the default is the current Region.

$sel:tableName:SchemaConfiguration', schemaConfiguration_tableName - Specifies the AWS Glue table that contains the column information that constitutes your data schema.

If the SchemaConfiguration request parameter is used as part of invoking the CreateDeliveryStream API, then the TableName property is required and its value must be specified.

$sel:roleARN:SchemaConfiguration', schemaConfiguration_roleARN - The role that Kinesis Data Firehose can use to access AWS Glue. This role must be in the same account you use for Kinesis Data Firehose. Cross-account roles aren't allowed.

If the SchemaConfiguration request parameter is used as part of invoking the CreateDeliveryStream API, then the RoleARN property is required and its value must be specified.

schemaConfiguration_versionId :: Lens' SchemaConfiguration (Maybe Text)

Specifies the table version for the output data schema. If you don't specify this version ID, or if you set it to LATEST, Kinesis Data Firehose uses the most recent version. This means that any updates to the table are automatically picked up.

schemaConfiguration_catalogId :: Lens' SchemaConfiguration (Maybe Text)

The ID of the AWS Glue Data Catalog. If you don't supply this, the AWS account ID is used by default.

schemaConfiguration_databaseName :: Lens' SchemaConfiguration (Maybe Text)

Specifies the name of the AWS Glue database that contains the schema for the output data.

If the SchemaConfiguration request parameter is used as part of invoking the CreateDeliveryStream API, then the DatabaseName property is required and its value must be specified.

schemaConfiguration_region :: Lens' SchemaConfiguration (Maybe Text)

If you don't specify an AWS Region, the default is the current Region.

schemaConfiguration_tableName :: Lens' SchemaConfiguration (Maybe Text)

Specifies the AWS Glue table that contains the column information that constitutes your data schema.

If the SchemaConfiguration request parameter is used as part of invoking the CreateDeliveryStream API, then the TableName property is required and its value must be specified.

schemaConfiguration_roleARN :: Lens' SchemaConfiguration (Maybe Text)

The role that Kinesis Data Firehose can use to access AWS Glue. This role must be in the same account you use for Kinesis Data Firehose. Cross-account roles aren't allowed.

If the SchemaConfiguration request parameter is used as part of invoking the CreateDeliveryStream API, then the RoleARN property is required and its value must be specified.