Network.AWS.Firehose.Types.SplunkDestinationDescription

Description

Documentation

data SplunkDestinationDescription

Describes a destination in Splunk.

See: newSplunkDestinationDescription smart constructor.

Constructors

SplunkDestinationDescription'

Fields

hECAcknowledgmentTimeoutInSeconds :: Maybe Natural

The amount of time that Kinesis Data Firehose waits to receive an acknowledgment from Splunk after it sends it data. At the end of the timeout period, Kinesis Data Firehose either tries to send the data again or considers it an error, based on your retry settings.

processingConfiguration :: Maybe ProcessingConfiguration

The data processing configuration.

cloudWatchLoggingOptions :: Maybe CloudWatchLoggingOptions

The Amazon CloudWatch logging options for your delivery stream.

hECEndpointType :: Maybe HECEndpointType

This type can be either "Raw" or "Event."

retryOptions :: Maybe SplunkRetryOptions

The retry behavior in case Kinesis Data Firehose is unable to deliver data to Splunk or if it doesn't receive an acknowledgment of receipt from Splunk.

s3BackupMode :: Maybe SplunkS3BackupMode

Defines how documents should be delivered to Amazon S3. When set to FailedDocumentsOnly, Kinesis Data Firehose writes any data that could not be indexed to the configured Amazon S3 destination. When set to AllDocuments, Kinesis Data Firehose delivers all incoming records to Amazon S3, and also writes failed documents to Amazon S3. Default value is FailedDocumentsOnly.

hECEndpoint :: Maybe Text

The HTTP Event Collector (HEC) endpoint to which Kinesis Data Firehose sends your data.

hECToken :: Maybe Text

A GUID you obtain from your Splunk cluster when you create a new HEC endpoint.

s3DestinationDescription :: Maybe S3DestinationDescription

The Amazon S3 destination.>

newSplunkDestinationDescription :: SplunkDestinationDescription

Create a value of SplunkDestinationDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hECAcknowledgmentTimeoutInSeconds:SplunkDestinationDescription', splunkDestinationDescription_hECAcknowledgmentTimeoutInSeconds - The amount of time that Kinesis Data Firehose waits to receive an acknowledgment from Splunk after it sends it data. At the end of the timeout period, Kinesis Data Firehose either tries to send the data again or considers it an error, based on your retry settings.

$sel:processingConfiguration:SplunkDestinationDescription', splunkDestinationDescription_processingConfiguration - The data processing configuration.

$sel:cloudWatchLoggingOptions:SplunkDestinationDescription', splunkDestinationDescription_cloudWatchLoggingOptions - The Amazon CloudWatch logging options for your delivery stream.

$sel:hECEndpointType:SplunkDestinationDescription', splunkDestinationDescription_hECEndpointType - This type can be either "Raw" or "Event."

$sel:retryOptions:SplunkDestinationDescription', splunkDestinationDescription_retryOptions - The retry behavior in case Kinesis Data Firehose is unable to deliver data to Splunk or if it doesn't receive an acknowledgment of receipt from Splunk.

$sel:s3BackupMode:SplunkDestinationDescription', splunkDestinationDescription_s3BackupMode - Defines how documents should be delivered to Amazon S3. When set to FailedDocumentsOnly, Kinesis Data Firehose writes any data that could not be indexed to the configured Amazon S3 destination. When set to AllDocuments, Kinesis Data Firehose delivers all incoming records to Amazon S3, and also writes failed documents to Amazon S3. Default value is FailedDocumentsOnly.

$sel:hECEndpoint:SplunkDestinationDescription', splunkDestinationDescription_hECEndpoint - The HTTP Event Collector (HEC) endpoint to which Kinesis Data Firehose sends your data.

$sel:hECToken:SplunkDestinationDescription', splunkDestinationDescription_hECToken - A GUID you obtain from your Splunk cluster when you create a new HEC endpoint.

$sel:s3DestinationDescription:SplunkDestinationDescription', splunkDestinationDescription_s3DestinationDescription - The Amazon S3 destination.>

splunkDestinationDescription_hECAcknowledgmentTimeoutInSeconds :: Lens' SplunkDestinationDescription (Maybe Natural)

The amount of time that Kinesis Data Firehose waits to receive an acknowledgment from Splunk after it sends it data. At the end of the timeout period, Kinesis Data Firehose either tries to send the data again or considers it an error, based on your retry settings.

splunkDestinationDescription_processingConfiguration :: Lens' SplunkDestinationDescription (Maybe ProcessingConfiguration)

The data processing configuration.

splunkDestinationDescription_cloudWatchLoggingOptions :: Lens' SplunkDestinationDescription (Maybe CloudWatchLoggingOptions)

The Amazon CloudWatch logging options for your delivery stream.

splunkDestinationDescription_hECEndpointType :: Lens' SplunkDestinationDescription (Maybe HECEndpointType)

This type can be either "Raw" or "Event."

splunkDestinationDescription_retryOptions :: Lens' SplunkDestinationDescription (Maybe SplunkRetryOptions)

The retry behavior in case Kinesis Data Firehose is unable to deliver data to Splunk or if it doesn't receive an acknowledgment of receipt from Splunk.

splunkDestinationDescription_s3BackupMode :: Lens' SplunkDestinationDescription (Maybe SplunkS3BackupMode)

Defines how documents should be delivered to Amazon S3. When set to FailedDocumentsOnly, Kinesis Data Firehose writes any data that could not be indexed to the configured Amazon S3 destination. When set to AllDocuments, Kinesis Data Firehose delivers all incoming records to Amazon S3, and also writes failed documents to Amazon S3. Default value is FailedDocumentsOnly.

splunkDestinationDescription_hECEndpoint :: Lens' SplunkDestinationDescription (Maybe Text)

The HTTP Event Collector (HEC) endpoint to which Kinesis Data Firehose sends your data.

splunkDestinationDescription_hECToken :: Lens' SplunkDestinationDescription (Maybe Text)

A GUID you obtain from your Splunk cluster when you create a new HEC endpoint.

splunkDestinationDescription_s3DestinationDescription :: Lens' SplunkDestinationDescription (Maybe S3DestinationDescription)

The Amazon S3 destination.>