Network.AWS.Firehose.Types.SplunkDestinationUpdate

Description

Documentation

data SplunkDestinationUpdate

Describes an update for a destination in Splunk.

See: newSplunkDestinationUpdate smart constructor.

Constructors

SplunkDestinationUpdate'

Fields

s3Update :: Maybe S3DestinationUpdate

Your update to the configuration of the backup Amazon S3 location.

hECAcknowledgmentTimeoutInSeconds :: Maybe Natural

The amount of time that Kinesis Data Firehose waits to receive an acknowledgment from Splunk after it sends data. At the end of the timeout period, Kinesis Data Firehose either tries to send the data again or considers it an error, based on your retry settings.

processingConfiguration :: Maybe ProcessingConfiguration

The data processing configuration.

cloudWatchLoggingOptions :: Maybe CloudWatchLoggingOptions

The Amazon CloudWatch logging options for your delivery stream.

hECEndpointType :: Maybe HECEndpointType

This type can be either "Raw" or "Event."

retryOptions :: Maybe SplunkRetryOptions

The retry behavior in case Kinesis Data Firehose is unable to deliver data to Splunk or if it doesn't receive an acknowledgment of receipt from Splunk.

s3BackupMode :: Maybe SplunkS3BackupMode

Specifies how you want Kinesis Data Firehose to back up documents to Amazon S3. When set to FailedDocumentsOnly, Kinesis Data Firehose writes any data that could not be indexed to the configured Amazon S3 destination. When set to AllEvents, Kinesis Data Firehose delivers all incoming records to Amazon S3, and also writes failed documents to Amazon S3. The default value is FailedEventsOnly.

You can update this backup mode from FailedEventsOnly to AllEvents. You can't update it from AllEvents to FailedEventsOnly.

hECEndpoint :: Maybe Text

The HTTP Event Collector (HEC) endpoint to which Kinesis Data Firehose sends your data.

hECToken :: Maybe Text

A GUID that you obtain from your Splunk cluster when you create a new HEC endpoint.

newSplunkDestinationUpdate :: SplunkDestinationUpdate

Create a value of SplunkDestinationUpdate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3Update:SplunkDestinationUpdate', splunkDestinationUpdate_s3Update - Your update to the configuration of the backup Amazon S3 location.

$sel:hECAcknowledgmentTimeoutInSeconds:SplunkDestinationUpdate', splunkDestinationUpdate_hECAcknowledgmentTimeoutInSeconds - The amount of time that Kinesis Data Firehose waits to receive an acknowledgment from Splunk after it sends data. At the end of the timeout period, Kinesis Data Firehose either tries to send the data again or considers it an error, based on your retry settings.

$sel:processingConfiguration:SplunkDestinationUpdate', splunkDestinationUpdate_processingConfiguration - The data processing configuration.

$sel:cloudWatchLoggingOptions:SplunkDestinationUpdate', splunkDestinationUpdate_cloudWatchLoggingOptions - The Amazon CloudWatch logging options for your delivery stream.

$sel:hECEndpointType:SplunkDestinationUpdate', splunkDestinationUpdate_hECEndpointType - This type can be either "Raw" or "Event."

$sel:retryOptions:SplunkDestinationUpdate', splunkDestinationUpdate_retryOptions - The retry behavior in case Kinesis Data Firehose is unable to deliver data to Splunk or if it doesn't receive an acknowledgment of receipt from Splunk.

$sel:s3BackupMode:SplunkDestinationUpdate', splunkDestinationUpdate_s3BackupMode - Specifies how you want Kinesis Data Firehose to back up documents to Amazon S3. When set to FailedDocumentsOnly, Kinesis Data Firehose writes any data that could not be indexed to the configured Amazon S3 destination. When set to AllEvents, Kinesis Data Firehose delivers all incoming records to Amazon S3, and also writes failed documents to Amazon S3. The default value is FailedEventsOnly.

You can update this backup mode from FailedEventsOnly to AllEvents. You can't update it from AllEvents to FailedEventsOnly.

$sel:hECEndpoint:SplunkDestinationUpdate', splunkDestinationUpdate_hECEndpoint - The HTTP Event Collector (HEC) endpoint to which Kinesis Data Firehose sends your data.

$sel:hECToken:SplunkDestinationUpdate', splunkDestinationUpdate_hECToken - A GUID that you obtain from your Splunk cluster when you create a new HEC endpoint.

splunkDestinationUpdate_s3Update :: Lens' SplunkDestinationUpdate (Maybe S3DestinationUpdate)

Your update to the configuration of the backup Amazon S3 location.

splunkDestinationUpdate_hECAcknowledgmentTimeoutInSeconds :: Lens' SplunkDestinationUpdate (Maybe Natural)

The amount of time that Kinesis Data Firehose waits to receive an acknowledgment from Splunk after it sends data. At the end of the timeout period, Kinesis Data Firehose either tries to send the data again or considers it an error, based on your retry settings.

splunkDestinationUpdate_processingConfiguration :: Lens' SplunkDestinationUpdate (Maybe ProcessingConfiguration)

The data processing configuration.

splunkDestinationUpdate_cloudWatchLoggingOptions :: Lens' SplunkDestinationUpdate (Maybe CloudWatchLoggingOptions)

The Amazon CloudWatch logging options for your delivery stream.

splunkDestinationUpdate_hECEndpointType :: Lens' SplunkDestinationUpdate (Maybe HECEndpointType)

This type can be either "Raw" or "Event."

splunkDestinationUpdate_retryOptions :: Lens' SplunkDestinationUpdate (Maybe SplunkRetryOptions)

The retry behavior in case Kinesis Data Firehose is unable to deliver data to Splunk or if it doesn't receive an acknowledgment of receipt from Splunk.

splunkDestinationUpdate_s3BackupMode :: Lens' SplunkDestinationUpdate (Maybe SplunkS3BackupMode)

Specifies how you want Kinesis Data Firehose to back up documents to Amazon S3. When set to FailedDocumentsOnly, Kinesis Data Firehose writes any data that could not be indexed to the configured Amazon S3 destination. When set to AllEvents, Kinesis Data Firehose delivers all incoming records to Amazon S3, and also writes failed documents to Amazon S3. The default value is FailedEventsOnly.

You can update this backup mode from FailedEventsOnly to AllEvents. You can't update it from AllEvents to FailedEventsOnly.

splunkDestinationUpdate_hECEndpoint :: Lens' SplunkDestinationUpdate (Maybe Text)

The HTTP Event Collector (HEC) endpoint to which Kinesis Data Firehose sends your data.

splunkDestinationUpdate_hECToken :: Lens' SplunkDestinationUpdate (Maybe Text)

A GUID that you obtain from your Splunk cluster when you create a new HEC endpoint.