Network.AWS.Forecast.CreateDataset

Description

Creates an Amazon Forecast dataset. The information about the dataset that you provide helps Forecast understand how to consume the data for model training. This includes the following:

DataFrequency - How frequently your historical time-series data is collected.

Domain and DatasetType - Each dataset has an associated dataset domain and a type within the domain. Amazon Forecast provides a list of predefined domains and types within each domain. For each unique dataset domain and type within the domain, Amazon Forecast requires your data to include a minimum set of predefined fields.

Schema - A schema specifies the fields in the dataset, including the field name and data type.

After creating a dataset, you import your training data into it and add the dataset to a dataset group. You use the dataset group to create a predictor. For more information, see howitworks-datasets-groups.

To get a list of all your datasets, use the ListDatasets operation.

For example Forecast datasets, see the Amazon Forecast Sample GitHub repository.

The Status of a dataset must be ACTIVE before you can import training data. Use the DescribeDataset operation to get the status.

Creating a Request

data CreateDataset

See: newCreateDataset smart constructor.

Constructors

CreateDataset'

Fields

dataFrequency :: Maybe Text

The frequency of data collection. This parameter is required for RELATED_TIME_SERIES datasets.

Valid intervals are Y (Year), M (Month), W (Week), D (Day), H (Hour), 30min (30 minutes), 15min (15 minutes), 10min (10 minutes), 5min (5 minutes), and 1min (1 minute). For example, "D" indicates every day and "15min" indicates every 15 minutes.

encryptionConfig :: Maybe EncryptionConfig

An AWS Key Management Service (KMS) key and the AWS Identity and Access Management (IAM) role that Amazon Forecast can assume to access the key.

tags :: Maybe [Tag]

The optional metadata that you apply to the dataset to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define.

The following basic restrictions apply to tags:

Maximum number of tags per resource - 50.

For each resource, each tag key must be unique, and each tag key can have only one value.

Maximum key length - 128 Unicode characters in UTF-8.

Maximum value length - 256 Unicode characters in UTF-8.

If your tagging schema is used across multiple services and resources, remember that other services may have restrictions on allowed characters. Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following characters: + - = . _ : / @.

Tag keys and values are case sensitive.

Do not use aws:, AWS:, or any upper or lowercase combination of such as a prefix for keys as it is reserved for AWS use. You cannot edit or delete tag keys with this prefix. Values can have this prefix. If a tag value has aws as its prefix but the key does not, then Forecast considers it to be a user tag and will count against the limit of 50 tags. Tags with only the key prefix of aws do not count against your tags per resource limit.

datasetName :: Text

A name for the dataset.

domain :: Domain

The domain associated with the dataset. When you add a dataset to a dataset group, this value and the value specified for the Domain parameter of the CreateDatasetGroup operation must match.

The Domain and DatasetType that you choose determine the fields that must be present in the training data that you import to the dataset. For example, if you choose the RETAIL domain and TARGET_TIME_SERIES as the DatasetType, Amazon Forecast requires item_id, timestamp, and demand fields to be present in your data. For more information, see howitworks-datasets-groups.

datasetType :: DatasetType

The dataset type. Valid values depend on the chosen Domain.

schema :: Schema

The schema for the dataset. The schema attributes and their order must match the fields in your data. The dataset Domain and DatasetType that you choose determine the minimum required fields in your training data. For information about the required fields for a specific dataset domain and type, see howitworks-domains-ds-types.

newCreateDataset

Arguments

:: Text
$sel:datasetName:CreateDataset'

-> Domain
$sel:domain:CreateDataset'

-> DatasetType
$sel:datasetType:CreateDataset'

-> Schema
$sel:schema:CreateDataset'

-> CreateDataset

Create a value of CreateDataset with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataFrequency:CreateDataset', createDataset_dataFrequency - The frequency of data collection. This parameter is required for RELATED_TIME_SERIES datasets.

Valid intervals are Y (Year), M (Month), W (Week), D (Day), H (Hour), 30min (30 minutes), 15min (15 minutes), 10min (10 minutes), 5min (5 minutes), and 1min (1 minute). For example, "D" indicates every day and "15min" indicates every 15 minutes.

$sel:encryptionConfig:CreateDataset', createDataset_encryptionConfig - An AWS Key Management Service (KMS) key and the AWS Identity and Access Management (IAM) role that Amazon Forecast can assume to access the key.

$sel:tags:CreateDataset', createDataset_tags - The optional metadata that you apply to the dataset to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define.

The following basic restrictions apply to tags:

Maximum number of tags per resource - 50.

For each resource, each tag key must be unique, and each tag key can have only one value.

Maximum key length - 128 Unicode characters in UTF-8.

Maximum value length - 256 Unicode characters in UTF-8.

If your tagging schema is used across multiple services and resources, remember that other services may have restrictions on allowed characters. Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following characters: + - = . _ : / @.

Tag keys and values are case sensitive.

Do not use aws:, AWS:, or any upper or lowercase combination of such as a prefix for keys as it is reserved for AWS use. You cannot edit or delete tag keys with this prefix. Values can have this prefix. If a tag value has aws as its prefix but the key does not, then Forecast considers it to be a user tag and will count against the limit of 50 tags. Tags with only the key prefix of aws do not count against your tags per resource limit.

$sel:datasetName:CreateDataset', createDataset_datasetName - A name for the dataset.

$sel:domain:CreateDataset', createDataset_domain - The domain associated with the dataset. When you add a dataset to a dataset group, this value and the value specified for the Domain parameter of the CreateDatasetGroup operation must match.

The Domain and DatasetType that you choose determine the fields that must be present in the training data that you import to the dataset. For example, if you choose the RETAIL domain and TARGET_TIME_SERIES as the DatasetType, Amazon Forecast requires item_id, timestamp, and demand fields to be present in your data. For more information, see howitworks-datasets-groups.

$sel:datasetType:CreateDataset', createDataset_datasetType - The dataset type. Valid values depend on the chosen Domain.

$sel:schema:CreateDataset', createDataset_schema - The schema for the dataset. The schema attributes and their order must match the fields in your data. The dataset Domain and DatasetType that you choose determine the minimum required fields in your training data. For information about the required fields for a specific dataset domain and type, see howitworks-domains-ds-types.

Request Lenses

createDataset_dataFrequency :: Lens' CreateDataset (Maybe Text)

The frequency of data collection. This parameter is required for RELATED_TIME_SERIES datasets.

Valid intervals are Y (Year), M (Month), W (Week), D (Day), H (Hour), 30min (30 minutes), 15min (15 minutes), 10min (10 minutes), 5min (5 minutes), and 1min (1 minute). For example, "D" indicates every day and "15min" indicates every 15 minutes.

createDataset_encryptionConfig :: Lens' CreateDataset (Maybe EncryptionConfig)

An AWS Key Management Service (KMS) key and the AWS Identity and Access Management (IAM) role that Amazon Forecast can assume to access the key.

createDataset_tags :: Lens' CreateDataset (Maybe [Tag])

The optional metadata that you apply to the dataset to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define.

The following basic restrictions apply to tags:

Maximum number of tags per resource - 50.

For each resource, each tag key must be unique, and each tag key can have only one value.

Maximum key length - 128 Unicode characters in UTF-8.

Maximum value length - 256 Unicode characters in UTF-8.

If your tagging schema is used across multiple services and resources, remember that other services may have restrictions on allowed characters. Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following characters: + - = . _ : / @.

Tag keys and values are case sensitive.

Do not use aws:, AWS:, or any upper or lowercase combination of such as a prefix for keys as it is reserved for AWS use. You cannot edit or delete tag keys with this prefix. Values can have this prefix. If a tag value has aws as its prefix but the key does not, then Forecast considers it to be a user tag and will count against the limit of 50 tags. Tags with only the key prefix of aws do not count against your tags per resource limit.

createDataset_datasetName :: Lens' CreateDataset Text

A name for the dataset.

createDataset_domain :: Lens' CreateDataset Domain

The domain associated with the dataset. When you add a dataset to a dataset group, this value and the value specified for the Domain parameter of the CreateDatasetGroup operation must match.

The Domain and DatasetType that you choose determine the fields that must be present in the training data that you import to the dataset. For example, if you choose the RETAIL domain and TARGET_TIME_SERIES as the DatasetType, Amazon Forecast requires item_id, timestamp, and demand fields to be present in your data. For more information, see howitworks-datasets-groups.

createDataset_datasetType :: Lens' CreateDataset DatasetType

The dataset type. Valid values depend on the chosen Domain.

createDataset_schema :: Lens' CreateDataset Schema

The schema for the dataset. The schema attributes and their order must match the fields in your data. The dataset Domain and DatasetType that you choose determine the minimum required fields in your training data. For information about the required fields for a specific dataset domain and type, see howitworks-domains-ds-types.

Destructuring the Response

data CreateDatasetResponse

See: newCreateDatasetResponse smart constructor.

Constructors

CreateDatasetResponse'

Fields

datasetArn :: Maybe Text

The Amazon Resource Name (ARN) of the dataset.

httpStatus :: Int

The response's http status code.

newCreateDatasetResponse

Arguments

:: Int
$sel:httpStatus:CreateDatasetResponse'

-> CreateDatasetResponse

Create a value of CreateDatasetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:datasetArn:CreateDatasetResponse', createDatasetResponse_datasetArn - The Amazon Resource Name (ARN) of the dataset.

$sel:httpStatus:CreateDatasetResponse', createDatasetResponse_httpStatus - The response's http status code.

Response Lenses

createDatasetResponse_datasetArn :: Lens' CreateDatasetResponse (Maybe Text)

The Amazon Resource Name (ARN) of the dataset.

createDatasetResponse_httpStatus :: Lens' CreateDatasetResponse Int

The response's http status code.