Network.AWS.Forecast.CreateDatasetGroup

Description

Creates a dataset group, which holds a collection of related datasets. You can add datasets to the dataset group when you create the dataset group, or later by using the UpdateDatasetGroup operation.

After creating a dataset group and adding datasets, you use the dataset group when you create a predictor. For more information, see howitworks-datasets-groups.

To get a list of all your datasets groups, use the ListDatasetGroups operation.

The Status of a dataset group must be ACTIVE before you can use the dataset group to create a predictor. To get the status, use the DescribeDatasetGroup operation.

Creating a Request

data CreateDatasetGroup

See: newCreateDatasetGroup smart constructor.

Constructors

CreateDatasetGroup'

Fields

datasetArns :: Maybe [Text]

An array of Amazon Resource Names (ARNs) of the datasets that you want to include in the dataset group.

tags :: Maybe [Tag]

The optional metadata that you apply to the dataset group to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define.

The following basic restrictions apply to tags:

Maximum number of tags per resource - 50.

For each resource, each tag key must be unique, and each tag key can have only one value.

Maximum key length - 128 Unicode characters in UTF-8.

Maximum value length - 256 Unicode characters in UTF-8.

If your tagging schema is used across multiple services and resources, remember that other services may have restrictions on allowed characters. Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following characters: + - = . _ : / @.

Tag keys and values are case sensitive.

Do not use aws:, AWS:, or any upper or lowercase combination of such as a prefix for keys as it is reserved for AWS use. You cannot edit or delete tag keys with this prefix. Values can have this prefix. If a tag value has aws as its prefix but the key does not, then Forecast considers it to be a user tag and will count against the limit of 50 tags. Tags with only the key prefix of aws do not count against your tags per resource limit.

datasetGroupName :: Text

A name for the dataset group.

domain :: Domain

The domain associated with the dataset group. When you add a dataset to a dataset group, this value and the value specified for the Domain parameter of the CreateDataset operation must match.

The Domain and DatasetType that you choose determine the fields that must be present in training data that you import to a dataset. For example, if you choose the RETAIL domain and TARGET_TIME_SERIES as the DatasetType, Amazon Forecast requires that item_id, timestamp, and demand fields are present in your data. For more information, see howitworks-datasets-groups.

newCreateDatasetGroup

Arguments

:: Text
$sel:datasetGroupName:CreateDatasetGroup'

-> Domain
$sel:domain:CreateDatasetGroup'

-> CreateDatasetGroup

Create a value of CreateDatasetGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:datasetArns:CreateDatasetGroup', createDatasetGroup_datasetArns - An array of Amazon Resource Names (ARNs) of the datasets that you want to include in the dataset group.

$sel:tags:CreateDatasetGroup', createDatasetGroup_tags - The optional metadata that you apply to the dataset group to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define.

The following basic restrictions apply to tags:

Maximum number of tags per resource - 50.

For each resource, each tag key must be unique, and each tag key can have only one value.

Maximum key length - 128 Unicode characters in UTF-8.

Maximum value length - 256 Unicode characters in UTF-8.

If your tagging schema is used across multiple services and resources, remember that other services may have restrictions on allowed characters. Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following characters: + - = . _ : / @.

Tag keys and values are case sensitive.

Do not use aws:, AWS:, or any upper or lowercase combination of such as a prefix for keys as it is reserved for AWS use. You cannot edit or delete tag keys with this prefix. Values can have this prefix. If a tag value has aws as its prefix but the key does not, then Forecast considers it to be a user tag and will count against the limit of 50 tags. Tags with only the key prefix of aws do not count against your tags per resource limit.

$sel:datasetGroupName:CreateDatasetGroup', createDatasetGroup_datasetGroupName - A name for the dataset group.

$sel:domain:CreateDatasetGroup', createDatasetGroup_domain - The domain associated with the dataset group. When you add a dataset to a dataset group, this value and the value specified for the Domain parameter of the CreateDataset operation must match.

The Domain and DatasetType that you choose determine the fields that must be present in training data that you import to a dataset. For example, if you choose the RETAIL domain and TARGET_TIME_SERIES as the DatasetType, Amazon Forecast requires that item_id, timestamp, and demand fields are present in your data. For more information, see howitworks-datasets-groups.

Request Lenses

createDatasetGroup_datasetArns :: Lens' CreateDatasetGroup (Maybe [Text])

An array of Amazon Resource Names (ARNs) of the datasets that you want to include in the dataset group.

createDatasetGroup_tags :: Lens' CreateDatasetGroup (Maybe [Tag])

The optional metadata that you apply to the dataset group to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define.

The following basic restrictions apply to tags:

Maximum number of tags per resource - 50.

For each resource, each tag key must be unique, and each tag key can have only one value.

Maximum key length - 128 Unicode characters in UTF-8.

Maximum value length - 256 Unicode characters in UTF-8.

If your tagging schema is used across multiple services and resources, remember that other services may have restrictions on allowed characters. Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following characters: + - = . _ : / @.

Tag keys and values are case sensitive.

Do not use aws:, AWS:, or any upper or lowercase combination of such as a prefix for keys as it is reserved for AWS use. You cannot edit or delete tag keys with this prefix. Values can have this prefix. If a tag value has aws as its prefix but the key does not, then Forecast considers it to be a user tag and will count against the limit of 50 tags. Tags with only the key prefix of aws do not count against your tags per resource limit.

createDatasetGroup_datasetGroupName :: Lens' CreateDatasetGroup Text

A name for the dataset group.

createDatasetGroup_domain :: Lens' CreateDatasetGroup Domain

The domain associated with the dataset group. When you add a dataset to a dataset group, this value and the value specified for the Domain parameter of the CreateDataset operation must match.

The Domain and DatasetType that you choose determine the fields that must be present in training data that you import to a dataset. For example, if you choose the RETAIL domain and TARGET_TIME_SERIES as the DatasetType, Amazon Forecast requires that item_id, timestamp, and demand fields are present in your data. For more information, see howitworks-datasets-groups.

Destructuring the Response

data CreateDatasetGroupResponse

See: newCreateDatasetGroupResponse smart constructor.

Constructors

CreateDatasetGroupResponse'

Fields

datasetGroupArn :: Maybe Text

The Amazon Resource Name (ARN) of the dataset group.

httpStatus :: Int

The response's http status code.

newCreateDatasetGroupResponse

Arguments

:: Int
$sel:httpStatus:CreateDatasetGroupResponse'

-> CreateDatasetGroupResponse

Create a value of CreateDatasetGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:datasetGroupArn:CreateDatasetGroupResponse', createDatasetGroupResponse_datasetGroupArn - The Amazon Resource Name (ARN) of the dataset group.

$sel:httpStatus:CreateDatasetGroupResponse', createDatasetGroupResponse_httpStatus - The response's http status code.

Response Lenses

createDatasetGroupResponse_datasetGroupArn :: Lens' CreateDatasetGroupResponse (Maybe Text)

The Amazon Resource Name (ARN) of the dataset group.

createDatasetGroupResponse_httpStatus :: Lens' CreateDatasetGroupResponse Int

The response's http status code.