Network.AWS.Forecast.CreateForecastExportJob

Description

Exports a forecast created by the CreateForecast operation to your Amazon Simple Storage Service (Amazon S3) bucket. The forecast file name will match the following conventions:

<ForecastExportJobName>_<ExportTimestamp>_<PartNumber>

where the <ExportTimestamp> component is in Java SimpleDateFormat (yyyy-MM-ddTHH-mm-ssZ).

You must specify a DataDestination object that includes an AWS Identity and Access Management (IAM) role that Amazon Forecast can assume to access the Amazon S3 bucket. For more information, see aws-forecast-iam-roles.

For more information, see howitworks-forecast.

To get a list of all your forecast export jobs, use the ListForecastExportJobs operation.

The Status of the forecast export job must be ACTIVE before you can access the forecast in your Amazon S3 bucket. To get the status, use the DescribeForecastExportJob operation.

Creating a Request

data CreateForecastExportJob

See: newCreateForecastExportJob smart constructor.

Constructors

CreateForecastExportJob'

Fields

tags :: Maybe [Tag]

The optional metadata that you apply to the forecast export job to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define.

The following basic restrictions apply to tags:

Maximum number of tags per resource - 50.

For each resource, each tag key must be unique, and each tag key can have only one value.

Maximum key length - 128 Unicode characters in UTF-8.

Maximum value length - 256 Unicode characters in UTF-8.

If your tagging schema is used across multiple services and resources, remember that other services may have restrictions on allowed characters. Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following characters: + - = . _ : / @.

Tag keys and values are case sensitive.

Do not use aws:, AWS:, or any upper or lowercase combination of such as a prefix for keys as it is reserved for AWS use. You cannot edit or delete tag keys with this prefix. Values can have this prefix. If a tag value has aws as its prefix but the key does not, then Forecast considers it to be a user tag and will count against the limit of 50 tags. Tags with only the key prefix of aws do not count against your tags per resource limit.

forecastExportJobName :: Text

The name for the forecast export job.

forecastArn :: Text

The Amazon Resource Name (ARN) of the forecast that you want to export.

destination :: DataDestination

The location where you want to save the forecast and an AWS Identity and Access Management (IAM) role that Amazon Forecast can assume to access the location. The forecast must be exported to an Amazon S3 bucket.

If encryption is used, Destination must include an AWS Key Management Service (KMS) key. The IAM role must allow Amazon Forecast permission to access the key.

newCreateForecastExportJob

Arguments

:: Text
$sel:forecastExportJobName:CreateForecastExportJob'

-> Text
$sel:forecastArn:CreateForecastExportJob'

-> DataDestination
$sel:destination:CreateForecastExportJob'

-> CreateForecastExportJob

Create a value of CreateForecastExportJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateForecastExportJob', createForecastExportJob_tags - The optional metadata that you apply to the forecast export job to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define.

The following basic restrictions apply to tags:

Maximum number of tags per resource - 50.

For each resource, each tag key must be unique, and each tag key can have only one value.

Maximum key length - 128 Unicode characters in UTF-8.

Maximum value length - 256 Unicode characters in UTF-8.

If your tagging schema is used across multiple services and resources, remember that other services may have restrictions on allowed characters. Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following characters: + - = . _ : / @.

Tag keys and values are case sensitive.

Do not use aws:, AWS:, or any upper or lowercase combination of such as a prefix for keys as it is reserved for AWS use. You cannot edit or delete tag keys with this prefix. Values can have this prefix. If a tag value has aws as its prefix but the key does not, then Forecast considers it to be a user tag and will count against the limit of 50 tags. Tags with only the key prefix of aws do not count against your tags per resource limit.

$sel:forecastExportJobName:CreateForecastExportJob', createForecastExportJob_forecastExportJobName - The name for the forecast export job.

$sel:forecastArn:CreateForecastExportJob', createForecastExportJob_forecastArn - The Amazon Resource Name (ARN) of the forecast that you want to export.

$sel:destination:CreateForecastExportJob', createForecastExportJob_destination - The location where you want to save the forecast and an AWS Identity and Access Management (IAM) role that Amazon Forecast can assume to access the location. The forecast must be exported to an Amazon S3 bucket.

If encryption is used, Destination must include an AWS Key Management Service (KMS) key. The IAM role must allow Amazon Forecast permission to access the key.

Request Lenses

createForecastExportJob_tags :: Lens' CreateForecastExportJob (Maybe [Tag])

The optional metadata that you apply to the forecast export job to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define.

The following basic restrictions apply to tags:

Maximum number of tags per resource - 50.

For each resource, each tag key must be unique, and each tag key can have only one value.

Maximum key length - 128 Unicode characters in UTF-8.

Maximum value length - 256 Unicode characters in UTF-8.

If your tagging schema is used across multiple services and resources, remember that other services may have restrictions on allowed characters. Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following characters: + - = . _ : / @.

Tag keys and values are case sensitive.

Do not use aws:, AWS:, or any upper or lowercase combination of such as a prefix for keys as it is reserved for AWS use. You cannot edit or delete tag keys with this prefix. Values can have this prefix. If a tag value has aws as its prefix but the key does not, then Forecast considers it to be a user tag and will count against the limit of 50 tags. Tags with only the key prefix of aws do not count against your tags per resource limit.

createForecastExportJob_forecastExportJobName :: Lens' CreateForecastExportJob Text

The name for the forecast export job.

createForecastExportJob_forecastArn :: Lens' CreateForecastExportJob Text

The Amazon Resource Name (ARN) of the forecast that you want to export.

createForecastExportJob_destination :: Lens' CreateForecastExportJob DataDestination

The location where you want to save the forecast and an AWS Identity and Access Management (IAM) role that Amazon Forecast can assume to access the location. The forecast must be exported to an Amazon S3 bucket.

If encryption is used, Destination must include an AWS Key Management Service (KMS) key. The IAM role must allow Amazon Forecast permission to access the key.

Destructuring the Response

data CreateForecastExportJobResponse

See: newCreateForecastExportJobResponse smart constructor.

Constructors

CreateForecastExportJobResponse'

Fields

forecastExportJobArn :: Maybe Text

The Amazon Resource Name (ARN) of the export job.

httpStatus :: Int

The response's http status code.

newCreateForecastExportJobResponse

Arguments

:: Int
$sel:httpStatus:CreateForecastExportJobResponse'

-> CreateForecastExportJobResponse

Create a value of CreateForecastExportJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:forecastExportJobArn:CreateForecastExportJobResponse', createForecastExportJobResponse_forecastExportJobArn - The Amazon Resource Name (ARN) of the export job.

$sel:httpStatus:CreateForecastExportJobResponse', createForecastExportJobResponse_httpStatus - The response's http status code.

Response Lenses

createForecastExportJobResponse_forecastExportJobArn :: Lens' CreateForecastExportJobResponse (Maybe Text)

The Amazon Resource Name (ARN) of the export job.

createForecastExportJobResponse_httpStatus :: Lens' CreateForecastExportJobResponse Int

The response's http status code.