Network.AWS.Forecast.CreatePredictor

Description

Creates an Amazon Forecast predictor.

In the request, provide a dataset group and either specify an algorithm or let Amazon Forecast choose an algorithm for you using AutoML. If you specify an algorithm, you also can override algorithm-specific hyperparameters.

Amazon Forecast uses the algorithm to train a predictor using the latest version of the datasets in the specified dataset group. You can then generate a forecast using the CreateForecast operation.

To see the evaluation metrics, use the GetAccuracyMetrics operation.

You can specify a featurization configuration to fill and aggregate the data fields in the TARGET_TIME_SERIES dataset to improve model training. For more information, see FeaturizationConfig.

For RELATED_TIME_SERIES datasets, CreatePredictor verifies that the DataFrequency specified when the dataset was created matches the ForecastFrequency. TARGET_TIME_SERIES datasets don't have this restriction. Amazon Forecast also verifies the delimiter and timestamp format. For more information, see howitworks-datasets-groups.

By default, predictors are trained and evaluated at the 0.1 (P10), 0.5 (P50), and 0.9 (P90) quantiles. You can choose custom forecast types to train and evaluate your predictor by setting the ForecastTypes.

AutoML

If you want Amazon Forecast to evaluate each algorithm and choose the one that minimizes the objective function, set PerformAutoML to true. The objective function is defined as the mean of the weighted losses over the forecast types. By default, these are the p10, p50, and p90 quantile losses. For more information, see EvaluationResult.

When AutoML is enabled, the following properties are disallowed:

AlgorithmArn

HPOConfig

PerformHPO

TrainingParameters

To get a list of all of your predictors, use the ListPredictors operation.

Before you can use the predictor to create a forecast, the Status of the predictor must be ACTIVE, signifying that training has completed. To get the status, use the DescribePredictor operation.

Creating a Request

data CreatePredictor

See: newCreatePredictor smart constructor.

Constructors

CreatePredictor'

Fields

performAutoML :: Maybe Bool

Whether to perform AutoML. When Amazon Forecast performs AutoML, it evaluates the algorithms it provides and chooses the best algorithm and configuration for your training dataset.

The default value is false. In this case, you are required to specify an algorithm.

Set PerformAutoML to true to have Amazon Forecast perform AutoML. This is a good option if you aren't sure which algorithm is suitable for your training data. In this case, PerformHPO must be false.

trainingParameters :: Maybe (HashMap Text Text)

The hyperparameters to override for model training. The hyperparameters that you can override are listed in the individual algorithms. For the list of supported algorithms, see aws-forecast-choosing-recipes.

algorithmArn :: Maybe Text

The Amazon Resource Name (ARN) of the algorithm to use for model training. Required if PerformAutoML is not set to true.

Supported algorithms:

arn:aws:forecast:::algorithm/ARIMA

arn:aws:forecast:::algorithm/CNN-QR

arn:aws:forecast:::algorithm/Deep_AR_Plus

arn:aws:forecast:::algorithm/ETS

arn:aws:forecast:::algorithm/NPTS

arn:aws:forecast:::algorithm/Prophet

hPOConfig :: Maybe HyperParameterTuningJobConfig

Provides hyperparameter override values for the algorithm. If you don't provide this parameter, Amazon Forecast uses default values. The individual algorithms specify which hyperparameters support hyperparameter optimization (HPO). For more information, see aws-forecast-choosing-recipes.

If you included the HPOConfig object, you must set PerformHPO to true.

optimizationMetric :: Maybe OptimizationMetric

The accuracy metric used to optimize the predictor.

autoMLOverrideStrategy :: Maybe AutoMLOverrideStrategy

The LatencyOptimized AutoML override strategy is only available in private beta. Contact AWS Support or your account manager to learn more about access privileges.

Used to overide the default AutoML strategy, which is to optimize predictor accuracy. To apply an AutoML strategy that minimizes training time, use LatencyOptimized.

This parameter is only valid for predictors trained using AutoML.

evaluationParameters :: Maybe EvaluationParameters

Used to override the default evaluation parameters of the specified algorithm. Amazon Forecast evaluates a predictor by splitting a dataset into training data and testing data. The evaluation parameters define how to perform the split and the number of iterations.

encryptionConfig :: Maybe EncryptionConfig

An AWS Key Management Service (KMS) key and the AWS Identity and Access Management (IAM) role that Amazon Forecast can assume to access the key.

forecastTypes :: Maybe (NonEmpty Text)

Specifies the forecast types used to train a predictor. You can specify up to five forecast types. Forecast types can be quantiles from 0.01 to 0.99, by increments of 0.01 or higher. You can also specify the mean forecast with mean.

The default value is ["0.10", "0.50", "0.9"].

performHPO :: Maybe Bool

Whether to perform hyperparameter optimization (HPO). HPO finds optimal hyperparameter values for your training data. The process of performing HPO is known as running a hyperparameter tuning job.

The default value is false. In this case, Amazon Forecast uses default hyperparameter values from the chosen algorithm.

To override the default values, set PerformHPO to true and, optionally, supply the HyperParameterTuningJobConfig object. The tuning job specifies a metric to optimize, which hyperparameters participate in tuning, and the valid range for each tunable hyperparameter. In this case, you are required to specify an algorithm and PerformAutoML must be false.

The following algorithms support HPO:

DeepAR+

CNN-QR

tags :: Maybe [Tag]

The optional metadata that you apply to the predictor to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define.

The following basic restrictions apply to tags:

Maximum number of tags per resource - 50.

For each resource, each tag key must be unique, and each tag key can have only one value.

Maximum key length - 128 Unicode characters in UTF-8.

Maximum value length - 256 Unicode characters in UTF-8.

If your tagging schema is used across multiple services and resources, remember that other services may have restrictions on allowed characters. Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following characters: + - = . _ : / @.

Tag keys and values are case sensitive.

Do not use aws:, AWS:, or any upper or lowercase combination of such as a prefix for keys as it is reserved for AWS use. You cannot edit or delete tag keys with this prefix. Values can have this prefix. If a tag value has aws as its prefix but the key does not, then Forecast considers it to be a user tag and will count against the limit of 50 tags. Tags with only the key prefix of aws do not count against your tags per resource limit.

predictorName :: Text

A name for the predictor.

forecastHorizon :: Int

Specifies the number of time-steps that the model is trained to predict. The forecast horizon is also called the prediction length.

For example, if you configure a dataset for daily data collection (using the DataFrequency parameter of the CreateDataset operation) and set the forecast horizon to 10, the model returns predictions for 10 days.

The maximum forecast horizon is the lesser of 500 time-steps or 1/3 of the TARGET_TIME_SERIES dataset length.

inputDataConfig :: InputDataConfig

Describes the dataset group that contains the data to use to train the predictor.

featurizationConfig :: FeaturizationConfig

The featurization configuration.

newCreatePredictor

Arguments

:: Text
$sel:predictorName:CreatePredictor'

-> Int
$sel:forecastHorizon:CreatePredictor'

-> InputDataConfig
$sel:inputDataConfig:CreatePredictor'

-> FeaturizationConfig
$sel:featurizationConfig:CreatePredictor'

-> CreatePredictor

Create a value of CreatePredictor with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:performAutoML:CreatePredictor', createPredictor_performAutoML - Whether to perform AutoML. When Amazon Forecast performs AutoML, it evaluates the algorithms it provides and chooses the best algorithm and configuration for your training dataset.

The default value is false. In this case, you are required to specify an algorithm.

Set PerformAutoML to true to have Amazon Forecast perform AutoML. This is a good option if you aren't sure which algorithm is suitable for your training data. In this case, PerformHPO must be false.

$sel:trainingParameters:CreatePredictor', createPredictor_trainingParameters - The hyperparameters to override for model training. The hyperparameters that you can override are listed in the individual algorithms. For the list of supported algorithms, see aws-forecast-choosing-recipes.

$sel:algorithmArn:CreatePredictor', createPredictor_algorithmArn - The Amazon Resource Name (ARN) of the algorithm to use for model training. Required if PerformAutoML is not set to true.

Supported algorithms:

arn:aws:forecast:::algorithm/ARIMA

arn:aws:forecast:::algorithm/CNN-QR

arn:aws:forecast:::algorithm/Deep_AR_Plus

arn:aws:forecast:::algorithm/ETS

arn:aws:forecast:::algorithm/NPTS

arn:aws:forecast:::algorithm/Prophet

$sel:hPOConfig:CreatePredictor', createPredictor_hPOConfig - Provides hyperparameter override values for the algorithm. If you don't provide this parameter, Amazon Forecast uses default values. The individual algorithms specify which hyperparameters support hyperparameter optimization (HPO). For more information, see aws-forecast-choosing-recipes.

If you included the HPOConfig object, you must set PerformHPO to true.

$sel:optimizationMetric:CreatePredictor', createPredictor_optimizationMetric - The accuracy metric used to optimize the predictor.

$sel:autoMLOverrideStrategy:CreatePredictor', createPredictor_autoMLOverrideStrategy - The LatencyOptimized AutoML override strategy is only available in private beta. Contact AWS Support or your account manager to learn more about access privileges.

Used to overide the default AutoML strategy, which is to optimize predictor accuracy. To apply an AutoML strategy that minimizes training time, use LatencyOptimized.

This parameter is only valid for predictors trained using AutoML.

$sel:evaluationParameters:CreatePredictor', createPredictor_evaluationParameters - Used to override the default evaluation parameters of the specified algorithm. Amazon Forecast evaluates a predictor by splitting a dataset into training data and testing data. The evaluation parameters define how to perform the split and the number of iterations.

$sel:encryptionConfig:CreatePredictor', createPredictor_encryptionConfig - An AWS Key Management Service (KMS) key and the AWS Identity and Access Management (IAM) role that Amazon Forecast can assume to access the key.

$sel:forecastTypes:CreatePredictor', createPredictor_forecastTypes - Specifies the forecast types used to train a predictor. You can specify up to five forecast types. Forecast types can be quantiles from 0.01 to 0.99, by increments of 0.01 or higher. You can also specify the mean forecast with mean.

The default value is ["0.10", "0.50", "0.9"].

$sel:performHPO:CreatePredictor', createPredictor_performHPO - Whether to perform hyperparameter optimization (HPO). HPO finds optimal hyperparameter values for your training data. The process of performing HPO is known as running a hyperparameter tuning job.

The default value is false. In this case, Amazon Forecast uses default hyperparameter values from the chosen algorithm.

To override the default values, set PerformHPO to true and, optionally, supply the HyperParameterTuningJobConfig object. The tuning job specifies a metric to optimize, which hyperparameters participate in tuning, and the valid range for each tunable hyperparameter. In this case, you are required to specify an algorithm and PerformAutoML must be false.

The following algorithms support HPO:

DeepAR+

CNN-QR

$sel:tags:CreatePredictor', createPredictor_tags - The optional metadata that you apply to the predictor to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define.

The following basic restrictions apply to tags:

Maximum number of tags per resource - 50.

For each resource, each tag key must be unique, and each tag key can have only one value.

Maximum key length - 128 Unicode characters in UTF-8.

Maximum value length - 256 Unicode characters in UTF-8.

If your tagging schema is used across multiple services and resources, remember that other services may have restrictions on allowed characters. Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following characters: + - = . _ : / @.

Tag keys and values are case sensitive.

Do not use aws:, AWS:, or any upper or lowercase combination of such as a prefix for keys as it is reserved for AWS use. You cannot edit or delete tag keys with this prefix. Values can have this prefix. If a tag value has aws as its prefix but the key does not, then Forecast considers it to be a user tag and will count against the limit of 50 tags. Tags with only the key prefix of aws do not count against your tags per resource limit.

$sel:predictorName:CreatePredictor', createPredictor_predictorName - A name for the predictor.

$sel:forecastHorizon:CreatePredictor', createPredictor_forecastHorizon - Specifies the number of time-steps that the model is trained to predict. The forecast horizon is also called the prediction length.

For example, if you configure a dataset for daily data collection (using the DataFrequency parameter of the CreateDataset operation) and set the forecast horizon to 10, the model returns predictions for 10 days.

The maximum forecast horizon is the lesser of 500 time-steps or 1/3 of the TARGET_TIME_SERIES dataset length.

$sel:inputDataConfig:CreatePredictor', createPredictor_inputDataConfig - Describes the dataset group that contains the data to use to train the predictor.

$sel:featurizationConfig:CreatePredictor', createPredictor_featurizationConfig - The featurization configuration.

Request Lenses

createPredictor_performAutoML :: Lens' CreatePredictor (Maybe Bool)

Whether to perform AutoML. When Amazon Forecast performs AutoML, it evaluates the algorithms it provides and chooses the best algorithm and configuration for your training dataset.

The default value is false. In this case, you are required to specify an algorithm.

Set PerformAutoML to true to have Amazon Forecast perform AutoML. This is a good option if you aren't sure which algorithm is suitable for your training data. In this case, PerformHPO must be false.

createPredictor_trainingParameters :: Lens' CreatePredictor (Maybe (HashMap Text Text))

The hyperparameters to override for model training. The hyperparameters that you can override are listed in the individual algorithms. For the list of supported algorithms, see aws-forecast-choosing-recipes.

createPredictor_algorithmArn :: Lens' CreatePredictor (Maybe Text)

The Amazon Resource Name (ARN) of the algorithm to use for model training. Required if PerformAutoML is not set to true.

Supported algorithms:

arn:aws:forecast:::algorithm/ARIMA

arn:aws:forecast:::algorithm/CNN-QR

arn:aws:forecast:::algorithm/Deep_AR_Plus

arn:aws:forecast:::algorithm/ETS

arn:aws:forecast:::algorithm/NPTS

arn:aws:forecast:::algorithm/Prophet

createPredictor_hPOConfig :: Lens' CreatePredictor (Maybe HyperParameterTuningJobConfig)

Provides hyperparameter override values for the algorithm. If you don't provide this parameter, Amazon Forecast uses default values. The individual algorithms specify which hyperparameters support hyperparameter optimization (HPO). For more information, see aws-forecast-choosing-recipes.

If you included the HPOConfig object, you must set PerformHPO to true.

createPredictor_optimizationMetric :: Lens' CreatePredictor (Maybe OptimizationMetric)

The accuracy metric used to optimize the predictor.

createPredictor_autoMLOverrideStrategy :: Lens' CreatePredictor (Maybe AutoMLOverrideStrategy)

The LatencyOptimized AutoML override strategy is only available in private beta. Contact AWS Support or your account manager to learn more about access privileges.

Used to overide the default AutoML strategy, which is to optimize predictor accuracy. To apply an AutoML strategy that minimizes training time, use LatencyOptimized.

This parameter is only valid for predictors trained using AutoML.

createPredictor_evaluationParameters :: Lens' CreatePredictor (Maybe EvaluationParameters)

Used to override the default evaluation parameters of the specified algorithm. Amazon Forecast evaluates a predictor by splitting a dataset into training data and testing data. The evaluation parameters define how to perform the split and the number of iterations.

createPredictor_encryptionConfig :: Lens' CreatePredictor (Maybe EncryptionConfig)

An AWS Key Management Service (KMS) key and the AWS Identity and Access Management (IAM) role that Amazon Forecast can assume to access the key.

createPredictor_forecastTypes :: Lens' CreatePredictor (Maybe (NonEmpty Text))

Specifies the forecast types used to train a predictor. You can specify up to five forecast types. Forecast types can be quantiles from 0.01 to 0.99, by increments of 0.01 or higher. You can also specify the mean forecast with mean.

The default value is ["0.10", "0.50", "0.9"].

createPredictor_performHPO :: Lens' CreatePredictor (Maybe Bool)

Whether to perform hyperparameter optimization (HPO). HPO finds optimal hyperparameter values for your training data. The process of performing HPO is known as running a hyperparameter tuning job.

The default value is false. In this case, Amazon Forecast uses default hyperparameter values from the chosen algorithm.

To override the default values, set PerformHPO to true and, optionally, supply the HyperParameterTuningJobConfig object. The tuning job specifies a metric to optimize, which hyperparameters participate in tuning, and the valid range for each tunable hyperparameter. In this case, you are required to specify an algorithm and PerformAutoML must be false.

The following algorithms support HPO:

DeepAR+

CNN-QR

createPredictor_tags :: Lens' CreatePredictor (Maybe [Tag])

The optional metadata that you apply to the predictor to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define.

The following basic restrictions apply to tags:

Maximum number of tags per resource - 50.

For each resource, each tag key must be unique, and each tag key can have only one value.

Maximum key length - 128 Unicode characters in UTF-8.

Maximum value length - 256 Unicode characters in UTF-8.

If your tagging schema is used across multiple services and resources, remember that other services may have restrictions on allowed characters. Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following characters: + - = . _ : / @.

Tag keys and values are case sensitive.

Do not use aws:, AWS:, or any upper or lowercase combination of such as a prefix for keys as it is reserved for AWS use. You cannot edit or delete tag keys with this prefix. Values can have this prefix. If a tag value has aws as its prefix but the key does not, then Forecast considers it to be a user tag and will count against the limit of 50 tags. Tags with only the key prefix of aws do not count against your tags per resource limit.

createPredictor_predictorName :: Lens' CreatePredictor Text

A name for the predictor.

createPredictor_forecastHorizon :: Lens' CreatePredictor Int

Specifies the number of time-steps that the model is trained to predict. The forecast horizon is also called the prediction length.

For example, if you configure a dataset for daily data collection (using the DataFrequency parameter of the CreateDataset operation) and set the forecast horizon to 10, the model returns predictions for 10 days.

The maximum forecast horizon is the lesser of 500 time-steps or 1/3 of the TARGET_TIME_SERIES dataset length.

createPredictor_inputDataConfig :: Lens' CreatePredictor InputDataConfig

Describes the dataset group that contains the data to use to train the predictor.

createPredictor_featurizationConfig :: Lens' CreatePredictor FeaturizationConfig

The featurization configuration.

Destructuring the Response

data CreatePredictorResponse

See: newCreatePredictorResponse smart constructor.

Constructors

CreatePredictorResponse'

Fields

predictorArn :: Maybe Text

The Amazon Resource Name (ARN) of the predictor.

httpStatus :: Int

The response's http status code.

newCreatePredictorResponse

Arguments

:: Int
$sel:httpStatus:CreatePredictorResponse'

-> CreatePredictorResponse

Create a value of CreatePredictorResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:predictorArn:CreatePredictorResponse', createPredictorResponse_predictorArn - The Amazon Resource Name (ARN) of the predictor.

$sel:httpStatus:CreatePredictorResponse', createPredictorResponse_httpStatus - The response's http status code.

Response Lenses

createPredictorResponse_predictorArn :: Lens' CreatePredictorResponse (Maybe Text)

The Amazon Resource Name (ARN) of the predictor.

createPredictorResponse_httpStatus :: Lens' CreatePredictorResponse Int

The response's http status code.