Network.AWS.Forecast.DescribeDatasetGroup

Description

Describes a dataset group created using the CreateDatasetGroup operation.

In addition to listing the parameters provided in the CreateDatasetGroup request, this operation includes the following properties:

DatasetArns - The datasets belonging to the group.

CreationTime

LastModificationTime

Status

Creating a Request

data DescribeDatasetGroup

See: newDescribeDatasetGroup smart constructor.

Constructors

DescribeDatasetGroup'

Fields

datasetGroupArn :: Text

The Amazon Resource Name (ARN) of the dataset group.

newDescribeDatasetGroup

Arguments

:: Text
$sel:datasetGroupArn:DescribeDatasetGroup'

-> DescribeDatasetGroup

Create a value of DescribeDatasetGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:datasetGroupArn:DescribeDatasetGroup', describeDatasetGroup_datasetGroupArn - The Amazon Resource Name (ARN) of the dataset group.

Request Lenses

describeDatasetGroup_datasetGroupArn :: Lens' DescribeDatasetGroup Text

The Amazon Resource Name (ARN) of the dataset group.

Destructuring the Response

data DescribeDatasetGroupResponse

See: newDescribeDatasetGroupResponse smart constructor.

Constructors

DescribeDatasetGroupResponse'

Fields

creationTime :: Maybe POSIX

When the dataset group was created.

datasetArns :: Maybe [Text]

An array of Amazon Resource Names (ARNs) of the datasets contained in the dataset group.

status :: Maybe Text

The status of the dataset group. States include:

ACTIVE

CREATE_PENDING, CREATE_IN_PROGRESS, CREATE_FAILED

DELETE_PENDING, DELETE_IN_PROGRESS, DELETE_FAILED

UPDATE_PENDING, UPDATE_IN_PROGRESS, UPDATE_FAILED

The UPDATE states apply when you call the UpdateDatasetGroup operation.

The Status of the dataset group must be ACTIVE before you can use the dataset group to create a predictor.

domain :: Maybe Domain

The domain associated with the dataset group.

datasetGroupName :: Maybe Text

The name of the dataset group.

datasetGroupArn :: Maybe Text

The ARN of the dataset group.

lastModificationTime :: Maybe POSIX

When the dataset group was created or last updated from a call to the UpdateDatasetGroup operation. While the dataset group is being updated, LastModificationTime is the current time of the DescribeDatasetGroup call.

httpStatus :: Int

The response's http status code.

newDescribeDatasetGroupResponse

Arguments

:: Int
$sel:httpStatus:DescribeDatasetGroupResponse'

-> DescribeDatasetGroupResponse

Create a value of DescribeDatasetGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:DescribeDatasetGroupResponse', describeDatasetGroupResponse_creationTime - When the dataset group was created.

$sel:datasetArns:DescribeDatasetGroupResponse', describeDatasetGroupResponse_datasetArns - An array of Amazon Resource Names (ARNs) of the datasets contained in the dataset group.

$sel:status:DescribeDatasetGroupResponse', describeDatasetGroupResponse_status - The status of the dataset group. States include:

ACTIVE

CREATE_PENDING, CREATE_IN_PROGRESS, CREATE_FAILED

DELETE_PENDING, DELETE_IN_PROGRESS, DELETE_FAILED

UPDATE_PENDING, UPDATE_IN_PROGRESS, UPDATE_FAILED

The UPDATE states apply when you call the UpdateDatasetGroup operation.

The Status of the dataset group must be ACTIVE before you can use the dataset group to create a predictor.

$sel:domain:DescribeDatasetGroupResponse', describeDatasetGroupResponse_domain - The domain associated with the dataset group.

$sel:datasetGroupName:DescribeDatasetGroupResponse', describeDatasetGroupResponse_datasetGroupName - The name of the dataset group.

$sel:datasetGroupArn:DescribeDatasetGroup', describeDatasetGroupResponse_datasetGroupArn - The ARN of the dataset group.

$sel:lastModificationTime:DescribeDatasetGroupResponse', describeDatasetGroupResponse_lastModificationTime - When the dataset group was created or last updated from a call to the UpdateDatasetGroup operation. While the dataset group is being updated, LastModificationTime is the current time of the DescribeDatasetGroup call.

$sel:httpStatus:DescribeDatasetGroupResponse', describeDatasetGroupResponse_httpStatus - The response's http status code.

Response Lenses

describeDatasetGroupResponse_creationTime :: Lens' DescribeDatasetGroupResponse (Maybe UTCTime)

When the dataset group was created.

describeDatasetGroupResponse_datasetArns :: Lens' DescribeDatasetGroupResponse (Maybe [Text])

An array of Amazon Resource Names (ARNs) of the datasets contained in the dataset group.

describeDatasetGroupResponse_status :: Lens' DescribeDatasetGroupResponse (Maybe Text)

The status of the dataset group. States include:

ACTIVE

CREATE_PENDING, CREATE_IN_PROGRESS, CREATE_FAILED

DELETE_PENDING, DELETE_IN_PROGRESS, DELETE_FAILED

UPDATE_PENDING, UPDATE_IN_PROGRESS, UPDATE_FAILED

The UPDATE states apply when you call the UpdateDatasetGroup operation.

The Status of the dataset group must be ACTIVE before you can use the dataset group to create a predictor.

describeDatasetGroupResponse_domain :: Lens' DescribeDatasetGroupResponse (Maybe Domain)

The domain associated with the dataset group.

describeDatasetGroupResponse_datasetGroupName :: Lens' DescribeDatasetGroupResponse (Maybe Text)

The name of the dataset group.

describeDatasetGroupResponse_datasetGroupArn :: Lens' DescribeDatasetGroupResponse (Maybe Text)

The ARN of the dataset group.

describeDatasetGroupResponse_lastModificationTime :: Lens' DescribeDatasetGroupResponse (Maybe UTCTime)

When the dataset group was created or last updated from a call to the UpdateDatasetGroup operation. While the dataset group is being updated, LastModificationTime is the current time of the DescribeDatasetGroup call.

describeDatasetGroupResponse_httpStatus :: Lens' DescribeDatasetGroupResponse Int

The response's http status code.