Network.AWS.Forecast.DescribeForecast

Description

Describes a forecast created using the CreateForecast operation.

In addition to listing the properties provided in the CreateForecast request, this operation lists the following properties:

DatasetGroupArn - The dataset group that provided the training data.

CreationTime

LastModificationTime

Status

Message - If an error occurred, information about the error.

Creating a Request

data DescribeForecast

See: newDescribeForecast smart constructor.

Constructors

DescribeForecast'

Fields

forecastArn :: Text

The Amazon Resource Name (ARN) of the forecast.

newDescribeForecast

Arguments

:: Text
$sel:forecastArn:DescribeForecast'

-> DescribeForecast

Create a value of DescribeForecast with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:forecastArn:DescribeForecast', describeForecast_forecastArn - The Amazon Resource Name (ARN) of the forecast.

Request Lenses

describeForecast_forecastArn :: Lens' DescribeForecast Text

The Amazon Resource Name (ARN) of the forecast.

Destructuring the Response

data DescribeForecastResponse

See: newDescribeForecastResponse smart constructor.

Constructors

DescribeForecastResponse'

Fields

creationTime :: Maybe POSIX

When the forecast creation task was created.

status :: Maybe Text

The status of the forecast. States include:

ACTIVE

CREATE_PENDING, CREATE_IN_PROGRESS, CREATE_FAILED

CREATE_STOPPING, CREATE_STOPPED

DELETE_PENDING, DELETE_IN_PROGRESS, DELETE_FAILED

The Status of the forecast must be ACTIVE before you can query or export the forecast.

predictorArn :: Maybe Text

The ARN of the predictor used to generate the forecast.

forecastArn :: Maybe Text

The forecast ARN as specified in the request.

estimatedTimeRemainingInMinutes :: Maybe Integer

The estimated time remaining in minutes for the forecast job to complete.

forecastName :: Maybe Text

The name of the forecast.

forecastTypes :: Maybe (NonEmpty Text)

The quantiles at which probabilistic forecasts were generated.

datasetGroupArn :: Maybe Text

The ARN of the dataset group that provided the data used to train the predictor.

message :: Maybe Text

If an error occurred, an informational message about the error.

lastModificationTime :: Maybe POSIX

The last time the resource was modified. The timestamp depends on the status of the job:

CREATE_PENDING - The CreationTime.

CREATE_IN_PROGRESS - The current timestamp.

CREATE_STOPPING - The current timestamp.

CREATE_STOPPED - When the job stopped.

ACTIVE or CREATE_FAILED - When the job finished or failed.

httpStatus :: Int

The response's http status code.

newDescribeForecastResponse

Arguments

:: Int
$sel:httpStatus:DescribeForecastResponse'

-> DescribeForecastResponse

Create a value of DescribeForecastResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:DescribeForecastResponse', describeForecastResponse_creationTime - When the forecast creation task was created.

$sel:status:DescribeForecastResponse', describeForecastResponse_status - The status of the forecast. States include:

ACTIVE

CREATE_PENDING, CREATE_IN_PROGRESS, CREATE_FAILED

CREATE_STOPPING, CREATE_STOPPED

DELETE_PENDING, DELETE_IN_PROGRESS, DELETE_FAILED

The Status of the forecast must be ACTIVE before you can query or export the forecast.

$sel:predictorArn:DescribeForecastResponse', describeForecastResponse_predictorArn - The ARN of the predictor used to generate the forecast.

$sel:forecastArn:DescribeForecast', describeForecastResponse_forecastArn - The forecast ARN as specified in the request.

$sel:estimatedTimeRemainingInMinutes:DescribeForecastResponse', describeForecastResponse_estimatedTimeRemainingInMinutes - The estimated time remaining in minutes for the forecast job to complete.

$sel:forecastName:DescribeForecastResponse', describeForecastResponse_forecastName - The name of the forecast.

$sel:forecastTypes:DescribeForecastResponse', describeForecastResponse_forecastTypes - The quantiles at which probabilistic forecasts were generated.

$sel:datasetGroupArn:DescribeForecastResponse', describeForecastResponse_datasetGroupArn - The ARN of the dataset group that provided the data used to train the predictor.

$sel:message:DescribeForecastResponse', describeForecastResponse_message - If an error occurred, an informational message about the error.

$sel:lastModificationTime:DescribeForecastResponse', describeForecastResponse_lastModificationTime - The last time the resource was modified. The timestamp depends on the status of the job:

CREATE_PENDING - The CreationTime.

CREATE_IN_PROGRESS - The current timestamp.

CREATE_STOPPING - The current timestamp.

CREATE_STOPPED - When the job stopped.

ACTIVE or CREATE_FAILED - When the job finished or failed.

$sel:httpStatus:DescribeForecastResponse', describeForecastResponse_httpStatus - The response's http status code.

Response Lenses

describeForecastResponse_creationTime :: Lens' DescribeForecastResponse (Maybe UTCTime)

When the forecast creation task was created.

describeForecastResponse_status :: Lens' DescribeForecastResponse (Maybe Text)

The status of the forecast. States include:

ACTIVE

CREATE_PENDING, CREATE_IN_PROGRESS, CREATE_FAILED

CREATE_STOPPING, CREATE_STOPPED

DELETE_PENDING, DELETE_IN_PROGRESS, DELETE_FAILED

The Status of the forecast must be ACTIVE before you can query or export the forecast.

describeForecastResponse_predictorArn :: Lens' DescribeForecastResponse (Maybe Text)

The ARN of the predictor used to generate the forecast.

describeForecastResponse_forecastArn :: Lens' DescribeForecastResponse (Maybe Text)

The forecast ARN as specified in the request.

describeForecastResponse_estimatedTimeRemainingInMinutes :: Lens' DescribeForecastResponse (Maybe Integer)

The estimated time remaining in minutes for the forecast job to complete.

describeForecastResponse_forecastName :: Lens' DescribeForecastResponse (Maybe Text)

The name of the forecast.

describeForecastResponse_forecastTypes :: Lens' DescribeForecastResponse (Maybe (NonEmpty Text))

The quantiles at which probabilistic forecasts were generated.

describeForecastResponse_datasetGroupArn :: Lens' DescribeForecastResponse (Maybe Text)

The ARN of the dataset group that provided the data used to train the predictor.

describeForecastResponse_message :: Lens' DescribeForecastResponse (Maybe Text)

If an error occurred, an informational message about the error.

describeForecastResponse_lastModificationTime :: Lens' DescribeForecastResponse (Maybe UTCTime)

The last time the resource was modified. The timestamp depends on the status of the job:

CREATE_PENDING - The CreationTime.

CREATE_IN_PROGRESS - The current timestamp.

CREATE_STOPPING - The current timestamp.

CREATE_STOPPED - When the job stopped.

ACTIVE or CREATE_FAILED - When the job finished or failed.

describeForecastResponse_httpStatus :: Lens' DescribeForecastResponse Int

The response's http status code.