Network.AWS.Forecast.ListDatasetImportJobs

Description

Returns a list of dataset import jobs created using the CreateDatasetImportJob operation. For each import job, this operation returns a summary of its properties, including its Amazon Resource Name (ARN). You can retrieve the complete set of properties by using the ARN with the DescribeDatasetImportJob operation. You can filter the list by providing an array of Filter objects.

This operation returns paginated results.

Creating a Request

data ListDatasetImportJobs

See: newListDatasetImportJobs smart constructor.

Constructors

ListDatasetImportJobs'

Fields

filters :: Maybe [Filter]

An array of filters. For each filter, you provide a condition and a match statement. The condition is either IS or IS_NOT, which specifies whether to include or exclude the datasets that match the statement from the list, respectively. The match statement consists of a key and a value.

Filter properties

Condition - The condition to apply. Valid values are IS and IS_NOT. To include the datasets that match the statement, specify IS. To exclude matching datasets, specify IS_NOT.

Key - The name of the parameter to filter on. Valid values are DatasetArn and Status.

Value - The value to match.

For example, to list all dataset import jobs whose status is ACTIVE, you specify the following filter:

"Filters": [ { "Condition": "IS", "Key": "Status", "Value": "ACTIVE" } ]

nextToken :: Maybe Text

If the result of the previous request was truncated, the response includes a NextToken. To retrieve the next set of results, use the token in the next request. Tokens expire after 24 hours.

maxResults :: Maybe Natural

The number of items to return in the response.

newListDatasetImportJobs :: ListDatasetImportJobs

Create a value of ListDatasetImportJobs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:ListDatasetImportJobs', listDatasetImportJobs_filters - An array of filters. For each filter, you provide a condition and a match statement. The condition is either IS or IS_NOT, which specifies whether to include or exclude the datasets that match the statement from the list, respectively. The match statement consists of a key and a value.

Filter properties

Condition - The condition to apply. Valid values are IS and IS_NOT. To include the datasets that match the statement, specify IS. To exclude matching datasets, specify IS_NOT.

Key - The name of the parameter to filter on. Valid values are DatasetArn and Status.

Value - The value to match.

For example, to list all dataset import jobs whose status is ACTIVE, you specify the following filter:

"Filters": [ { "Condition": "IS", "Key": "Status", "Value": "ACTIVE" } ]

$sel:nextToken:ListDatasetImportJobs', listDatasetImportJobs_nextToken - If the result of the previous request was truncated, the response includes a NextToken. To retrieve the next set of results, use the token in the next request. Tokens expire after 24 hours.

$sel:maxResults:ListDatasetImportJobs', listDatasetImportJobs_maxResults - The number of items to return in the response.

Request Lenses

listDatasetImportJobs_filters :: Lens' ListDatasetImportJobs (Maybe [Filter])

An array of filters. For each filter, you provide a condition and a match statement. The condition is either IS or IS_NOT, which specifies whether to include or exclude the datasets that match the statement from the list, respectively. The match statement consists of a key and a value.

Filter properties

Condition - The condition to apply. Valid values are IS and IS_NOT. To include the datasets that match the statement, specify IS. To exclude matching datasets, specify IS_NOT.

Key - The name of the parameter to filter on. Valid values are DatasetArn and Status.

Value - The value to match.

For example, to list all dataset import jobs whose status is ACTIVE, you specify the following filter:

"Filters": [ { "Condition": "IS", "Key": "Status", "Value": "ACTIVE" } ]

listDatasetImportJobs_nextToken :: Lens' ListDatasetImportJobs (Maybe Text)

If the result of the previous request was truncated, the response includes a NextToken. To retrieve the next set of results, use the token in the next request. Tokens expire after 24 hours.

listDatasetImportJobs_maxResults :: Lens' ListDatasetImportJobs (Maybe Natural)

The number of items to return in the response.

Destructuring the Response

data ListDatasetImportJobsResponse

See: newListDatasetImportJobsResponse smart constructor.

Constructors

ListDatasetImportJobsResponse'

Fields

datasetImportJobs :: Maybe [DatasetImportJobSummary]

An array of objects that summarize each dataset import job's properties.

nextToken :: Maybe Text

If the response is truncated, Amazon Forecast returns this token. To retrieve the next set of results, use the token in the next request.

httpStatus :: Int

The response's http status code.

newListDatasetImportJobsResponse

Arguments

:: Int
$sel:httpStatus:ListDatasetImportJobsResponse'

-> ListDatasetImportJobsResponse

Create a value of ListDatasetImportJobsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:datasetImportJobs:ListDatasetImportJobsResponse', listDatasetImportJobsResponse_datasetImportJobs - An array of objects that summarize each dataset import job's properties.

$sel:nextToken:ListDatasetImportJobs', listDatasetImportJobsResponse_nextToken - If the response is truncated, Amazon Forecast returns this token. To retrieve the next set of results, use the token in the next request.

$sel:httpStatus:ListDatasetImportJobsResponse', listDatasetImportJobsResponse_httpStatus - The response's http status code.

Response Lenses

listDatasetImportJobsResponse_datasetImportJobs :: Lens' ListDatasetImportJobsResponse (Maybe [DatasetImportJobSummary])

An array of objects that summarize each dataset import job's properties.

listDatasetImportJobsResponse_nextToken :: Lens' ListDatasetImportJobsResponse (Maybe Text)

If the response is truncated, Amazon Forecast returns this token. To retrieve the next set of results, use the token in the next request.

listDatasetImportJobsResponse_httpStatus :: Lens' ListDatasetImportJobsResponse Int

The response's http status code.