Network.AWS.Forecast.ListForecasts

Description

Returns a list of forecasts created using the CreateForecast operation. For each forecast, this operation returns a summary of its properties, including its Amazon Resource Name (ARN). To retrieve the complete set of properties, specify the ARN with the DescribeForecast operation. You can filter the list using an array of Filter objects.

This operation returns paginated results.

Creating a Request

data ListForecasts

See: newListForecasts smart constructor.

Constructors

ListForecasts'

Fields

filters :: Maybe [Filter]

An array of filters. For each filter, you provide a condition and a match statement. The condition is either IS or IS_NOT, which specifies whether to include or exclude the forecasts that match the statement from the list, respectively. The match statement consists of a key and a value.

Filter properties

Condition - The condition to apply. Valid values are IS and IS_NOT. To include the forecasts that match the statement, specify IS. To exclude matching forecasts, specify IS_NOT.

Key - The name of the parameter to filter on. Valid values are DatasetGroupArn, PredictorArn, and Status.

Value - The value to match.

For example, to list all forecasts whose status is not ACTIVE, you would specify:

"Filters": [ { "Condition": "IS_NOT", "Key": "Status", "Value": "ACTIVE" } ]

nextToken :: Maybe Text

If the result of the previous request was truncated, the response includes a NextToken. To retrieve the next set of results, use the token in the next request. Tokens expire after 24 hours.

maxResults :: Maybe Natural

The number of items to return in the response.

newListForecasts :: ListForecasts

Create a value of ListForecasts with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:ListForecasts', listForecasts_filters - An array of filters. For each filter, you provide a condition and a match statement. The condition is either IS or IS_NOT, which specifies whether to include or exclude the forecasts that match the statement from the list, respectively. The match statement consists of a key and a value.

Filter properties

Condition - The condition to apply. Valid values are IS and IS_NOT. To include the forecasts that match the statement, specify IS. To exclude matching forecasts, specify IS_NOT.

Key - The name of the parameter to filter on. Valid values are DatasetGroupArn, PredictorArn, and Status.

Value - The value to match.

For example, to list all forecasts whose status is not ACTIVE, you would specify:

"Filters": [ { "Condition": "IS_NOT", "Key": "Status", "Value": "ACTIVE" } ]

$sel:nextToken:ListForecasts', listForecasts_nextToken - If the result of the previous request was truncated, the response includes a NextToken. To retrieve the next set of results, use the token in the next request. Tokens expire after 24 hours.

$sel:maxResults:ListForecasts', listForecasts_maxResults - The number of items to return in the response.

Request Lenses

listForecasts_filters :: Lens' ListForecasts (Maybe [Filter])

An array of filters. For each filter, you provide a condition and a match statement. The condition is either IS or IS_NOT, which specifies whether to include or exclude the forecasts that match the statement from the list, respectively. The match statement consists of a key and a value.

Filter properties

Condition - The condition to apply. Valid values are IS and IS_NOT. To include the forecasts that match the statement, specify IS. To exclude matching forecasts, specify IS_NOT.

Key - The name of the parameter to filter on. Valid values are DatasetGroupArn, PredictorArn, and Status.

Value - The value to match.

For example, to list all forecasts whose status is not ACTIVE, you would specify:

"Filters": [ { "Condition": "IS_NOT", "Key": "Status", "Value": "ACTIVE" } ]

listForecasts_nextToken :: Lens' ListForecasts (Maybe Text)

If the result of the previous request was truncated, the response includes a NextToken. To retrieve the next set of results, use the token in the next request. Tokens expire after 24 hours.

listForecasts_maxResults :: Lens' ListForecasts (Maybe Natural)

The number of items to return in the response.

Destructuring the Response

data ListForecastsResponse

See: newListForecastsResponse smart constructor.

Constructors

ListForecastsResponse'

Fields

forecasts :: Maybe [ForecastSummary]

An array of objects that summarize each forecast's properties.

nextToken :: Maybe Text

If the response is truncated, Amazon Forecast returns this token. To retrieve the next set of results, use the token in the next request.

httpStatus :: Int

The response's http status code.

newListForecastsResponse

Arguments

:: Int
$sel:httpStatus:ListForecastsResponse'

-> ListForecastsResponse

Create a value of ListForecastsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:forecasts:ListForecastsResponse', listForecastsResponse_forecasts - An array of objects that summarize each forecast's properties.

$sel:nextToken:ListForecasts', listForecastsResponse_nextToken - If the response is truncated, Amazon Forecast returns this token. To retrieve the next set of results, use the token in the next request.

$sel:httpStatus:ListForecastsResponse', listForecastsResponse_httpStatus - The response's http status code.

Response Lenses

listForecastsResponse_forecasts :: Lens' ListForecastsResponse (Maybe [ForecastSummary])

An array of objects that summarize each forecast's properties.

listForecastsResponse_nextToken :: Lens' ListForecastsResponse (Maybe Text)

If the response is truncated, Amazon Forecast returns this token. To retrieve the next set of results, use the token in the next request.

listForecastsResponse_httpStatus :: Lens' ListForecastsResponse Int

The response's http status code.