Network.AWS.Forecast.ListPredictorBacktestExportJobs

Description

Returns a list of predictor backtest export jobs created using the CreatePredictorBacktestExportJob operation. This operation returns a summary for each backtest export job. You can filter the list using an array of Filter objects.

To retrieve the complete set of properties for a particular backtest export job, use the ARN with the DescribePredictorBacktestExportJob operation.

This operation returns paginated results.

Creating a Request

data ListPredictorBacktestExportJobs

See: newListPredictorBacktestExportJobs smart constructor.

Constructors

ListPredictorBacktestExportJobs'

Fields

filters :: Maybe [Filter]

An array of filters. For each filter, provide a condition and a match statement. The condition is either IS or IS_NOT, which specifies whether to include or exclude the predictor backtest export jobs that match the statement from the list. The match statement consists of a key and a value.

Filter properties

Condition - The condition to apply. Valid values are IS and IS_NOT. To include the predictor backtest export jobs that match the statement, specify IS. To exclude matching predictor backtest export jobs, specify IS_NOT.

Key - The name of the parameter to filter on. Valid values are PredictorArn and Status.

Value - The value to match.

nextToken :: Maybe Text

If the result of the previous request was truncated, the response includes a NextToken. To retrieve the next set of results, use the token in the next request. Tokens expire after 24 hours.

maxResults :: Maybe Natural

The number of items to return in the response.

newListPredictorBacktestExportJobs :: ListPredictorBacktestExportJobs

Create a value of ListPredictorBacktestExportJobs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:ListPredictorBacktestExportJobs', listPredictorBacktestExportJobs_filters - An array of filters. For each filter, provide a condition and a match statement. The condition is either IS or IS_NOT, which specifies whether to include or exclude the predictor backtest export jobs that match the statement from the list. The match statement consists of a key and a value.

Filter properties

Condition - The condition to apply. Valid values are IS and IS_NOT. To include the predictor backtest export jobs that match the statement, specify IS. To exclude matching predictor backtest export jobs, specify IS_NOT.

Key - The name of the parameter to filter on. Valid values are PredictorArn and Status.

Value - The value to match.

$sel:nextToken:ListPredictorBacktestExportJobs', listPredictorBacktestExportJobs_nextToken - If the result of the previous request was truncated, the response includes a NextToken. To retrieve the next set of results, use the token in the next request. Tokens expire after 24 hours.

$sel:maxResults:ListPredictorBacktestExportJobs', listPredictorBacktestExportJobs_maxResults - The number of items to return in the response.

Request Lenses

listPredictorBacktestExportJobs_filters :: Lens' ListPredictorBacktestExportJobs (Maybe [Filter])

An array of filters. For each filter, provide a condition and a match statement. The condition is either IS or IS_NOT, which specifies whether to include or exclude the predictor backtest export jobs that match the statement from the list. The match statement consists of a key and a value.

Filter properties

Condition - The condition to apply. Valid values are IS and IS_NOT. To include the predictor backtest export jobs that match the statement, specify IS. To exclude matching predictor backtest export jobs, specify IS_NOT.

Key - The name of the parameter to filter on. Valid values are PredictorArn and Status.

Value - The value to match.

listPredictorBacktestExportJobs_nextToken :: Lens' ListPredictorBacktestExportJobs (Maybe Text)

If the result of the previous request was truncated, the response includes a NextToken. To retrieve the next set of results, use the token in the next request. Tokens expire after 24 hours.

listPredictorBacktestExportJobs_maxResults :: Lens' ListPredictorBacktestExportJobs (Maybe Natural)

The number of items to return in the response.

Destructuring the Response

data ListPredictorBacktestExportJobsResponse

See: newListPredictorBacktestExportJobsResponse smart constructor.

Constructors

ListPredictorBacktestExportJobsResponse'

Fields

nextToken :: Maybe Text

Returns this token if the response is truncated. To retrieve the next set of results, use the token in the next request.

predictorBacktestExportJobs :: Maybe [PredictorBacktestExportJobSummary]

An array of objects that summarize the properties of each predictor backtest export job.

httpStatus :: Int

The response's http status code.

newListPredictorBacktestExportJobsResponse

Arguments

:: Int
$sel:httpStatus:ListPredictorBacktestExportJobsResponse'

-> ListPredictorBacktestExportJobsResponse

Create a value of ListPredictorBacktestExportJobsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListPredictorBacktestExportJobs', listPredictorBacktestExportJobsResponse_nextToken - Returns this token if the response is truncated. To retrieve the next set of results, use the token in the next request.

$sel:predictorBacktestExportJobs:ListPredictorBacktestExportJobsResponse', listPredictorBacktestExportJobsResponse_predictorBacktestExportJobs - An array of objects that summarize the properties of each predictor backtest export job.

$sel:httpStatus:ListPredictorBacktestExportJobsResponse', listPredictorBacktestExportJobsResponse_httpStatus - The response's http status code.

Response Lenses

listPredictorBacktestExportJobsResponse_nextToken :: Lens' ListPredictorBacktestExportJobsResponse (Maybe Text)

Returns this token if the response is truncated. To retrieve the next set of results, use the token in the next request.

listPredictorBacktestExportJobsResponse_predictorBacktestExportJobs :: Lens' ListPredictorBacktestExportJobsResponse (Maybe [PredictorBacktestExportJobSummary])

An array of objects that summarize the properties of each predictor backtest export job.

listPredictorBacktestExportJobsResponse_httpStatus :: Lens' ListPredictorBacktestExportJobsResponse Int

The response's http status code.