Network.AWS.Forecast.Types.DatasetImportJobSummary

Description

Documentation

data DatasetImportJobSummary

Provides a summary of the dataset import job properties used in the ListDatasetImportJobs operation. To get the complete set of properties, call the DescribeDatasetImportJob operation, and provide the DatasetImportJobArn.

See: newDatasetImportJobSummary smart constructor.

Constructors

DatasetImportJobSummary'

Fields

creationTime :: Maybe POSIX

When the dataset import job was created.

status :: Maybe Text

The status of the dataset import job. States include:

ACTIVE

CREATE_PENDING, CREATE_IN_PROGRESS, CREATE_FAILED

DELETE_PENDING, DELETE_IN_PROGRESS, DELETE_FAILED

CREATE_STOPPING, CREATE_STOPPED

datasetImportJobName :: Maybe Text

The name of the dataset import job.

dataSource :: Maybe DataSource

The location of the training data to import and an AWS Identity and Access Management (IAM) role that Amazon Forecast can assume to access the data. The training data must be stored in an Amazon S3 bucket.

If encryption is used, DataSource includes an AWS Key Management Service (KMS) key.

datasetImportJobArn :: Maybe Text

The Amazon Resource Name (ARN) of the dataset import job.

message :: Maybe Text

If an error occurred, an informational message about the error.

lastModificationTime :: Maybe POSIX

The last time the resource was modified. The timestamp depends on the status of the job:

CREATE_PENDING - The CreationTime.

CREATE_IN_PROGRESS - The current timestamp.

CREATE_STOPPING - The current timestamp.

CREATE_STOPPED - When the job stopped.

ACTIVE or CREATE_FAILED - When the job finished or failed.

newDatasetImportJobSummary :: DatasetImportJobSummary

Create a value of DatasetImportJobSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:DatasetImportJobSummary', datasetImportJobSummary_creationTime - When the dataset import job was created.

$sel:status:DatasetImportJobSummary', datasetImportJobSummary_status - The status of the dataset import job. States include:

ACTIVE

CREATE_PENDING, CREATE_IN_PROGRESS, CREATE_FAILED

DELETE_PENDING, DELETE_IN_PROGRESS, DELETE_FAILED

CREATE_STOPPING, CREATE_STOPPED

$sel:datasetImportJobName:DatasetImportJobSummary', datasetImportJobSummary_datasetImportJobName - The name of the dataset import job.

$sel:dataSource:DatasetImportJobSummary', datasetImportJobSummary_dataSource - The location of the training data to import and an AWS Identity and Access Management (IAM) role that Amazon Forecast can assume to access the data. The training data must be stored in an Amazon S3 bucket.

If encryption is used, DataSource includes an AWS Key Management Service (KMS) key.

$sel:datasetImportJobArn:DatasetImportJobSummary', datasetImportJobSummary_datasetImportJobArn - The Amazon Resource Name (ARN) of the dataset import job.

$sel:message:DatasetImportJobSummary', datasetImportJobSummary_message - If an error occurred, an informational message about the error.

$sel:lastModificationTime:DatasetImportJobSummary', datasetImportJobSummary_lastModificationTime - The last time the resource was modified. The timestamp depends on the status of the job:

CREATE_PENDING - The CreationTime.

CREATE_IN_PROGRESS - The current timestamp.

CREATE_STOPPING - The current timestamp.

CREATE_STOPPED - When the job stopped.

ACTIVE or CREATE_FAILED - When the job finished or failed.

datasetImportJobSummary_creationTime :: Lens' DatasetImportJobSummary (Maybe UTCTime)

When the dataset import job was created.

datasetImportJobSummary_status :: Lens' DatasetImportJobSummary (Maybe Text)

The status of the dataset import job. States include:

ACTIVE

CREATE_PENDING, CREATE_IN_PROGRESS, CREATE_FAILED

DELETE_PENDING, DELETE_IN_PROGRESS, DELETE_FAILED

CREATE_STOPPING, CREATE_STOPPED

datasetImportJobSummary_datasetImportJobName :: Lens' DatasetImportJobSummary (Maybe Text)

The name of the dataset import job.

datasetImportJobSummary_dataSource :: Lens' DatasetImportJobSummary (Maybe DataSource)

The location of the training data to import and an AWS Identity and Access Management (IAM) role that Amazon Forecast can assume to access the data. The training data must be stored in an Amazon S3 bucket.

If encryption is used, DataSource includes an AWS Key Management Service (KMS) key.

datasetImportJobSummary_datasetImportJobArn :: Lens' DatasetImportJobSummary (Maybe Text)

The Amazon Resource Name (ARN) of the dataset import job.

datasetImportJobSummary_message :: Lens' DatasetImportJobSummary (Maybe Text)

If an error occurred, an informational message about the error.

datasetImportJobSummary_lastModificationTime :: Lens' DatasetImportJobSummary (Maybe UTCTime)

The last time the resource was modified. The timestamp depends on the status of the job:

CREATE_PENDING - The CreationTime.

CREATE_IN_PROGRESS - The current timestamp.

CREATE_STOPPING - The current timestamp.

CREATE_STOPPED - When the job stopped.

ACTIVE or CREATE_FAILED - When the job finished or failed.