Network.AWS.Forecast.Types.ErrorMetric

Description

Documentation

data ErrorMetric

Provides detailed error metrics to evaluate the performance of a predictor. This object is part of the Metrics object.

See: newErrorMetric smart constructor.

Constructors

ErrorMetric'

Fields

mase :: Maybe Double

The Mean Absolute Scaled Error (MASE)

wape :: Maybe Double

The weighted absolute percentage error (WAPE).

mape :: Maybe Double

The Mean Absolute Percentage Error (MAPE)

rmse :: Maybe Double

The root-mean-square error (RMSE).

forecastType :: Maybe Text

The Forecast type used to compute WAPE, MAPE, MASE, and RMSE.

newErrorMetric :: ErrorMetric

Create a value of ErrorMetric with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:mase:ErrorMetric', errorMetric_mase - The Mean Absolute Scaled Error (MASE)

$sel:wape:ErrorMetric', errorMetric_wape - The weighted absolute percentage error (WAPE).

$sel:mape:ErrorMetric', errorMetric_mape - The Mean Absolute Percentage Error (MAPE)

$sel:rmse:ErrorMetric', errorMetric_rmse - The root-mean-square error (RMSE).

$sel:forecastType:ErrorMetric', errorMetric_forecastType - The Forecast type used to compute WAPE, MAPE, MASE, and RMSE.

errorMetric_mase :: Lens' ErrorMetric (Maybe Double)

The Mean Absolute Scaled Error (MASE)

errorMetric_wape :: Lens' ErrorMetric (Maybe Double)

The weighted absolute percentage error (WAPE).

errorMetric_mape :: Lens' ErrorMetric (Maybe Double)

The Mean Absolute Percentage Error (MAPE)

errorMetric_rmse :: Lens' ErrorMetric (Maybe Double)

The root-mean-square error (RMSE).

errorMetric_forecastType :: Lens' ErrorMetric (Maybe Text)

The Forecast type used to compute WAPE, MAPE, MASE, and RMSE.