Network.AWS.Forecast.Types.ForecastExportJobSummary

Description

Documentation

data ForecastExportJobSummary

Provides a summary of the forecast export job properties used in the ListForecastExportJobs operation. To get the complete set of properties, call the DescribeForecastExportJob operation, and provide the listed ForecastExportJobArn.

See: newForecastExportJobSummary smart constructor.

Constructors

ForecastExportJobSummary'

Fields

creationTime :: Maybe POSIX

When the forecast export job was created.

status :: Maybe Text

The status of the forecast export job. States include:

ACTIVE

CREATE_PENDING, CREATE_IN_PROGRESS, CREATE_FAILED

CREATE_STOPPING, CREATE_STOPPED

DELETE_PENDING, DELETE_IN_PROGRESS, DELETE_FAILED

The Status of the forecast export job must be ACTIVE before you can access the forecast in your S3 bucket.

destination :: Maybe DataDestination

The path to the Amazon Simple Storage Service (Amazon S3) bucket where the forecast is exported.

forecastExportJobArn :: Maybe Text

The Amazon Resource Name (ARN) of the forecast export job.

forecastExportJobName :: Maybe Text

The name of the forecast export job.

message :: Maybe Text

If an error occurred, an informational message about the error.

lastModificationTime :: Maybe POSIX

The last time the resource was modified. The timestamp depends on the status of the job:

CREATE_PENDING - The CreationTime.

CREATE_IN_PROGRESS - The current timestamp.

CREATE_STOPPING - The current timestamp.

CREATE_STOPPED - When the job stopped.

ACTIVE or CREATE_FAILED - When the job finished or failed.

newForecastExportJobSummary :: ForecastExportJobSummary

Create a value of ForecastExportJobSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:ForecastExportJobSummary', forecastExportJobSummary_creationTime - When the forecast export job was created.

$sel:status:ForecastExportJobSummary', forecastExportJobSummary_status - The status of the forecast export job. States include:

ACTIVE

CREATE_PENDING, CREATE_IN_PROGRESS, CREATE_FAILED

CREATE_STOPPING, CREATE_STOPPED

DELETE_PENDING, DELETE_IN_PROGRESS, DELETE_FAILED

The Status of the forecast export job must be ACTIVE before you can access the forecast in your S3 bucket.

$sel:destination:ForecastExportJobSummary', forecastExportJobSummary_destination - The path to the Amazon Simple Storage Service (Amazon S3) bucket where the forecast is exported.

$sel:forecastExportJobArn:ForecastExportJobSummary', forecastExportJobSummary_forecastExportJobArn - The Amazon Resource Name (ARN) of the forecast export job.

$sel:forecastExportJobName:ForecastExportJobSummary', forecastExportJobSummary_forecastExportJobName - The name of the forecast export job.

$sel:message:ForecastExportJobSummary', forecastExportJobSummary_message - If an error occurred, an informational message about the error.

$sel:lastModificationTime:ForecastExportJobSummary', forecastExportJobSummary_lastModificationTime - The last time the resource was modified. The timestamp depends on the status of the job:

CREATE_PENDING - The CreationTime.

CREATE_IN_PROGRESS - The current timestamp.

CREATE_STOPPING - The current timestamp.

CREATE_STOPPED - When the job stopped.

ACTIVE or CREATE_FAILED - When the job finished or failed.

forecastExportJobSummary_creationTime :: Lens' ForecastExportJobSummary (Maybe UTCTime)

When the forecast export job was created.

forecastExportJobSummary_status :: Lens' ForecastExportJobSummary (Maybe Text)

The status of the forecast export job. States include:

ACTIVE

CREATE_PENDING, CREATE_IN_PROGRESS, CREATE_FAILED

CREATE_STOPPING, CREATE_STOPPED

DELETE_PENDING, DELETE_IN_PROGRESS, DELETE_FAILED

The Status of the forecast export job must be ACTIVE before you can access the forecast in your S3 bucket.

forecastExportJobSummary_destination :: Lens' ForecastExportJobSummary (Maybe DataDestination)

The path to the Amazon Simple Storage Service (Amazon S3) bucket where the forecast is exported.

forecastExportJobSummary_forecastExportJobArn :: Lens' ForecastExportJobSummary (Maybe Text)

The Amazon Resource Name (ARN) of the forecast export job.

forecastExportJobSummary_forecastExportJobName :: Lens' ForecastExportJobSummary (Maybe Text)

The name of the forecast export job.

forecastExportJobSummary_message :: Lens' ForecastExportJobSummary (Maybe Text)

If an error occurred, an informational message about the error.

forecastExportJobSummary_lastModificationTime :: Lens' ForecastExportJobSummary (Maybe UTCTime)

The last time the resource was modified. The timestamp depends on the status of the job:

CREATE_PENDING - The CreationTime.

CREATE_IN_PROGRESS - The current timestamp.

CREATE_STOPPING - The current timestamp.

CREATE_STOPPED - When the job stopped.

ACTIVE or CREATE_FAILED - When the job finished or failed.