Network.AWS.Forecast.Types.IntegerParameterRange

Description

Documentation

data IntegerParameterRange

Specifies an integer hyperparameter and it's range of tunable values. This object is part of the ParameterRanges object.

See: newIntegerParameterRange smart constructor.

Constructors

IntegerParameterRange'

Fields

scalingType :: Maybe ScalingType

The scale that hyperparameter tuning uses to search the hyperparameter range. Valid values:

Auto

Amazon Forecast hyperparameter tuning chooses the best scale for the hyperparameter.

Linear

Hyperparameter tuning searches the values in the hyperparameter range by using a linear scale.

Logarithmic

Hyperparameter tuning searches the values in the hyperparameter range by using a logarithmic scale.
Logarithmic scaling works only for ranges that have values greater than 0.

ReverseLogarithmic

Not supported for IntegerParameterRange.
Reverse logarithmic scaling works only for ranges that are entirely within the range 0 <= x < 1.0.

For information about choosing a hyperparameter scale, see Hyperparameter Scaling. One of the following values:

name :: Text

The name of the hyperparameter to tune.

maxValue :: Int

The maximum tunable value of the hyperparameter.

minValue :: Int

The minimum tunable value of the hyperparameter.

newIntegerParameterRange

Arguments

:: Text
$sel:name:IntegerParameterRange'

-> Int
$sel:maxValue:IntegerParameterRange'

-> Int
$sel:minValue:IntegerParameterRange'

-> IntegerParameterRange

Create a value of IntegerParameterRange with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:scalingType:IntegerParameterRange', integerParameterRange_scalingType - The scale that hyperparameter tuning uses to search the hyperparameter range. Valid values:

Auto

Amazon Forecast hyperparameter tuning chooses the best scale for the hyperparameter.

Linear

Hyperparameter tuning searches the values in the hyperparameter range by using a linear scale.

Logarithmic

Hyperparameter tuning searches the values in the hyperparameter range by using a logarithmic scale.
Logarithmic scaling works only for ranges that have values greater than 0.

ReverseLogarithmic

Not supported for IntegerParameterRange.
Reverse logarithmic scaling works only for ranges that are entirely within the range 0 <= x < 1.0.

For information about choosing a hyperparameter scale, see Hyperparameter Scaling. One of the following values:

$sel:name:IntegerParameterRange', integerParameterRange_name - The name of the hyperparameter to tune.

$sel:maxValue:IntegerParameterRange', integerParameterRange_maxValue - The maximum tunable value of the hyperparameter.

$sel:minValue:IntegerParameterRange', integerParameterRange_minValue - The minimum tunable value of the hyperparameter.

integerParameterRange_scalingType :: Lens' IntegerParameterRange (Maybe ScalingType)

The scale that hyperparameter tuning uses to search the hyperparameter range. Valid values:

Auto

Amazon Forecast hyperparameter tuning chooses the best scale for the hyperparameter.

Linear

Hyperparameter tuning searches the values in the hyperparameter range by using a linear scale.

Logarithmic

Hyperparameter tuning searches the values in the hyperparameter range by using a logarithmic scale.
Logarithmic scaling works only for ranges that have values greater than 0.

ReverseLogarithmic

Not supported for IntegerParameterRange.
Reverse logarithmic scaling works only for ranges that are entirely within the range 0 <= x < 1.0.

For information about choosing a hyperparameter scale, see Hyperparameter Scaling. One of the following values:

integerParameterRange_name :: Lens' IntegerParameterRange Text

The name of the hyperparameter to tune.

integerParameterRange_maxValue :: Lens' IntegerParameterRange Int

The maximum tunable value of the hyperparameter.

integerParameterRange_minValue :: Lens' IntegerParameterRange Int

The minimum tunable value of the hyperparameter.