Network.AWS.Forecast.Types.PredictorBacktestExportJobSummary

Description

Documentation

data PredictorBacktestExportJobSummary

Provides a summary of the predictor backtest export job properties used in the ListPredictorBacktestExportJobs operation. To get a complete set of properties, call the DescribePredictorBacktestExportJob operation, and provide the listed PredictorBacktestExportJobArn.

See: newPredictorBacktestExportJobSummary smart constructor.

Constructors

PredictorBacktestExportJobSummary'

Fields

creationTime :: Maybe POSIX

When the predictor backtest export job was created.

status :: Maybe Text

The status of the predictor backtest export job. States include:

ACTIVE

CREATE_PENDING, CREATE_IN_PROGRESS, CREATE_FAILED

CREATE_STOPPING, CREATE_STOPPED

DELETE_PENDING, DELETE_IN_PROGRESS, DELETE_FAILED

destination :: Maybe DataDestination

predictorBacktestExportJobArn :: Maybe Text

The Amazon Resource Name (ARN) of the predictor backtest export job.

message :: Maybe Text

Information about any errors that may have occurred during the backtest export.

predictorBacktestExportJobName :: Maybe Text

The name of the predictor backtest export job.

lastModificationTime :: Maybe POSIX

The last time the resource was modified. The timestamp depends on the status of the job:

CREATE_PENDING - The CreationTime.

CREATE_IN_PROGRESS - The current timestamp.

CREATE_STOPPING - The current timestamp.

CREATE_STOPPED - When the job stopped.

ACTIVE or CREATE_FAILED - When the job finished or failed.

newPredictorBacktestExportJobSummary :: PredictorBacktestExportJobSummary

Create a value of PredictorBacktestExportJobSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:PredictorBacktestExportJobSummary', predictorBacktestExportJobSummary_creationTime - When the predictor backtest export job was created.

$sel:status:PredictorBacktestExportJobSummary', predictorBacktestExportJobSummary_status - The status of the predictor backtest export job. States include:

ACTIVE

CREATE_PENDING, CREATE_IN_PROGRESS, CREATE_FAILED

CREATE_STOPPING, CREATE_STOPPED

DELETE_PENDING, DELETE_IN_PROGRESS, DELETE_FAILED

$sel:destination:PredictorBacktestExportJobSummary', predictorBacktestExportJobSummary_destination - Undocumented member.

$sel:predictorBacktestExportJobArn:PredictorBacktestExportJobSummary', predictorBacktestExportJobSummary_predictorBacktestExportJobArn - The Amazon Resource Name (ARN) of the predictor backtest export job.

$sel:message:PredictorBacktestExportJobSummary', predictorBacktestExportJobSummary_message - Information about any errors that may have occurred during the backtest export.

$sel:predictorBacktestExportJobName:PredictorBacktestExportJobSummary', predictorBacktestExportJobSummary_predictorBacktestExportJobName - The name of the predictor backtest export job.

$sel:lastModificationTime:PredictorBacktestExportJobSummary', predictorBacktestExportJobSummary_lastModificationTime - The last time the resource was modified. The timestamp depends on the status of the job:

CREATE_PENDING - The CreationTime.

CREATE_IN_PROGRESS - The current timestamp.

CREATE_STOPPING - The current timestamp.

CREATE_STOPPED - When the job stopped.

ACTIVE or CREATE_FAILED - When the job finished or failed.

predictorBacktestExportJobSummary_creationTime :: Lens' PredictorBacktestExportJobSummary (Maybe UTCTime)

When the predictor backtest export job was created.

predictorBacktestExportJobSummary_status :: Lens' PredictorBacktestExportJobSummary (Maybe Text)

The status of the predictor backtest export job. States include:

ACTIVE

CREATE_PENDING, CREATE_IN_PROGRESS, CREATE_FAILED

CREATE_STOPPING, CREATE_STOPPED

DELETE_PENDING, DELETE_IN_PROGRESS, DELETE_FAILED

predictorBacktestExportJobSummary_destination :: Lens' PredictorBacktestExportJobSummary (Maybe DataDestination)

Undocumented member.

predictorBacktestExportJobSummary_predictorBacktestExportJobArn :: Lens' PredictorBacktestExportJobSummary (Maybe Text)

The Amazon Resource Name (ARN) of the predictor backtest export job.

predictorBacktestExportJobSummary_message :: Lens' PredictorBacktestExportJobSummary (Maybe Text)

Information about any errors that may have occurred during the backtest export.

predictorBacktestExportJobSummary_predictorBacktestExportJobName :: Lens' PredictorBacktestExportJobSummary (Maybe Text)

The name of the predictor backtest export job.

predictorBacktestExportJobSummary_lastModificationTime :: Lens' PredictorBacktestExportJobSummary (Maybe UTCTime)

The last time the resource was modified. The timestamp depends on the status of the job:

CREATE_PENDING - The CreationTime.

CREATE_IN_PROGRESS - The current timestamp.

CREATE_STOPPING - The current timestamp.

CREATE_STOPPED - When the job stopped.

ACTIVE or CREATE_FAILED - When the job finished or failed.