Network.AWS.Forecast.Types.S3Config

Description

Documentation

data S3Config

The path to the file(s) in an Amazon Simple Storage Service (Amazon S3) bucket, and an AWS Identity and Access Management (IAM) role that Amazon Forecast can assume to access the file(s). Optionally, includes an AWS Key Management Service (KMS) key. This object is part of the DataSource object that is submitted in the CreateDatasetImportJob request, and part of the DataDestination object.

See: newS3Config smart constructor.

Constructors

S3Config'

Fields

kmsKeyArn :: Maybe Text

The Amazon Resource Name (ARN) of an AWS Key Management Service (KMS) key.

path :: Text

The path to an Amazon Simple Storage Service (Amazon S3) bucket or file(s) in an Amazon S3 bucket.

roleArn :: Text

The ARN of the AWS Identity and Access Management (IAM) role that Amazon Forecast can assume to access the Amazon S3 bucket or files. If you provide a value for the KMSKeyArn key, the role must allow access to the key.

Passing a role across AWS accounts is not allowed. If you pass a role that isn't in your account, you get an InvalidInputException error.

newS3Config

Arguments

:: Text
$sel:path:S3Config'

-> Text
$sel:roleArn:S3Config'

-> S3Config

Create a value of S3Config with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:kmsKeyArn:S3Config', s3Config_kmsKeyArn - The Amazon Resource Name (ARN) of an AWS Key Management Service (KMS) key.

$sel:path:S3Config', s3Config_path - The path to an Amazon Simple Storage Service (Amazon S3) bucket or file(s) in an Amazon S3 bucket.

$sel:roleArn:S3Config', s3Config_roleArn - The ARN of the AWS Identity and Access Management (IAM) role that Amazon Forecast can assume to access the Amazon S3 bucket or files. If you provide a value for the KMSKeyArn key, the role must allow access to the key.

Passing a role across AWS accounts is not allowed. If you pass a role that isn't in your account, you get an InvalidInputException error.

s3Config_kmsKeyArn :: Lens' S3Config (Maybe Text)

The Amazon Resource Name (ARN) of an AWS Key Management Service (KMS) key.

s3Config_path :: Lens' S3Config Text

The path to an Amazon Simple Storage Service (Amazon S3) bucket or file(s) in an Amazon S3 bucket.

s3Config_roleArn :: Lens' S3Config Text

The ARN of the AWS Identity and Access Management (IAM) role that Amazon Forecast can assume to access the Amazon S3 bucket or files. If you provide a value for the KMSKeyArn key, the role must allow access to the key.

Passing a role across AWS accounts is not allowed. If you pass a role that isn't in your account, you get an InvalidInputException error.