Network.AWS.Forecast.Types.Statistics

Description

Documentation

data Statistics

Provides statistics for each data field imported into to an Amazon Forecast dataset with the CreateDatasetImportJob operation.

See: newStatistics smart constructor.

Constructors

Statistics'

Fields

max :: Maybe Text

For a numeric field, the maximum value in the field.

countNullLong :: Maybe Integer

The number of null values in the field. CountNullLong is used instead of CountNull if the value is greater than 2,147,483,647.

countNan :: Maybe Int

The number of NAN (not a number) values in the field. If the response value is -1, refer to CountNanLong.

countNanLong :: Maybe Integer

The number of NAN (not a number) values in the field. CountNanLong is used instead of CountNan if the value is greater than 2,147,483,647.

avg :: Maybe Double

For a numeric field, the average value in the field.

countNull :: Maybe Int

The number of null values in the field. If the response value is -1, refer to CountNullLong.

count :: Maybe Int

The number of values in the field. If the response value is -1, refer to CountLong.

countLong :: Maybe Integer

The number of values in the field. CountLong is used instead of Count if the value is greater than 2,147,483,647.

stddev :: Maybe Double

For a numeric field, the standard deviation.

min :: Maybe Text

For a numeric field, the minimum value in the field.

countDistinctLong :: Maybe Integer

The number of distinct values in the field. CountDistinctLong is used instead of CountDistinct if the value is greater than 2,147,483,647.

countDistinct :: Maybe Int

The number of distinct values in the field. If the response value is -1, refer to CountDistinctLong.

newStatistics :: Statistics

Create a value of Statistics with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:max:Statistics', statistics_max - For a numeric field, the maximum value in the field.

$sel:countNullLong:Statistics', statistics_countNullLong - The number of null values in the field. CountNullLong is used instead of CountNull if the value is greater than 2,147,483,647.

$sel:countNan:Statistics', statistics_countNan - The number of NAN (not a number) values in the field. If the response value is -1, refer to CountNanLong.

$sel:countNanLong:Statistics', statistics_countNanLong - The number of NAN (not a number) values in the field. CountNanLong is used instead of CountNan if the value is greater than 2,147,483,647.

$sel:avg:Statistics', statistics_avg - For a numeric field, the average value in the field.

$sel:countNull:Statistics', statistics_countNull - The number of null values in the field. If the response value is -1, refer to CountNullLong.

$sel:count:Statistics', statistics_count - The number of values in the field. If the response value is -1, refer to CountLong.

$sel:countLong:Statistics', statistics_countLong - The number of values in the field. CountLong is used instead of Count if the value is greater than 2,147,483,647.

$sel:stddev:Statistics', statistics_stddev - For a numeric field, the standard deviation.

$sel:min:Statistics', statistics_min - For a numeric field, the minimum value in the field.

$sel:countDistinctLong:Statistics', statistics_countDistinctLong - The number of distinct values in the field. CountDistinctLong is used instead of CountDistinct if the value is greater than 2,147,483,647.

$sel:countDistinct:Statistics', statistics_countDistinct - The number of distinct values in the field. If the response value is -1, refer to CountDistinctLong.

statistics_max :: Lens' Statistics (Maybe Text)

For a numeric field, the maximum value in the field.

statistics_countNullLong :: Lens' Statistics (Maybe Integer)

The number of null values in the field. CountNullLong is used instead of CountNull if the value is greater than 2,147,483,647.

statistics_countNan :: Lens' Statistics (Maybe Int)

The number of NAN (not a number) values in the field. If the response value is -1, refer to CountNanLong.

statistics_countNanLong :: Lens' Statistics (Maybe Integer)

The number of NAN (not a number) values in the field. CountNanLong is used instead of CountNan if the value is greater than 2,147,483,647.

statistics_avg :: Lens' Statistics (Maybe Double)

For a numeric field, the average value in the field.

statistics_countNull :: Lens' Statistics (Maybe Int)

The number of null values in the field. If the response value is -1, refer to CountNullLong.

statistics_count :: Lens' Statistics (Maybe Int)

The number of values in the field. If the response value is -1, refer to CountLong.

statistics_countLong :: Lens' Statistics (Maybe Integer)

The number of values in the field. CountLong is used instead of Count if the value is greater than 2,147,483,647.

statistics_stddev :: Lens' Statistics (Maybe Double)

For a numeric field, the standard deviation.

statistics_min :: Lens' Statistics (Maybe Text)

For a numeric field, the minimum value in the field.

statistics_countDistinctLong :: Lens' Statistics (Maybe Integer)

The number of distinct values in the field. CountDistinctLong is used instead of CountDistinct if the value is greater than 2,147,483,647.

statistics_countDistinct :: Lens' Statistics (Maybe Int)

The number of distinct values in the field. If the response value is -1, refer to CountDistinctLong.