Network.AWS.ForecastQuery

Description

Derived from API version 2018-06-26 of the AWS service descriptions, licensed under Apache 2.0.

Provides APIs for creating and managing Amazon Forecast resources.

Service Configuration

defaultService :: Service

API version 2018-06-26 of the Amazon Forecast Query Service SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by ForecastQuery.

InvalidNextTokenException

_InvalidNextTokenException :: AsError a => Getting (First ServiceError) a ServiceError

The token is not valid. Tokens expire after 24 hours.

InvalidInputException

_InvalidInputException :: AsError a => Getting (First ServiceError) a ServiceError

The value is invalid or is too long.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

We can't find that resource. Check the information that you've provided and try again.

LimitExceededException

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The limit on the number of requests per second has been exceeded.

ResourceInUseException

_ResourceInUseException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource is in use.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

QueryForecast

data QueryForecast

See: newQueryForecast smart constructor.

Constructors

QueryForecast' (Maybe Text) (Maybe Text) (Maybe Text) Text (HashMap Text Text)

newQueryForecast

Arguments

:: Text
$sel:forecastArn:QueryForecast'

-> QueryForecast

Create a value of QueryForecast with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:endDate:QueryForecast', queryForecast_endDate - The end date for the forecast. Specify the date using this format: yyyy-MM-dd'T'HH:mm:ss (ISO 8601 format). For example, 2015-01-01T20:00:00.

$sel:startDate:QueryForecast', queryForecast_startDate - The start date for the forecast. Specify the date using this format: yyyy-MM-dd'T'HH:mm:ss (ISO 8601 format). For example, 2015-01-01T08:00:00.

$sel:nextToken:QueryForecast', queryForecast_nextToken - If the result of the previous request was truncated, the response includes a NextToken. To retrieve the next set of results, use the token in the next request. Tokens expire after 24 hours.

$sel:forecastArn:QueryForecast', queryForecast_forecastArn - The Amazon Resource Name (ARN) of the forecast to query.

$sel:filters:QueryForecast', queryForecast_filters - The filtering criteria to apply when retrieving the forecast. For example, to get the forecast for client_21 in the electricity usage dataset, specify the following:

{"item_id" : "client_21"}

To get the full forecast, use the CreateForecastExportJob operation.

data QueryForecastResponse

See: newQueryForecastResponse smart constructor.

Constructors

QueryForecastResponse' (Maybe Forecast) Int

newQueryForecastResponse

Arguments

:: Int
$sel:httpStatus:QueryForecastResponse'

-> QueryForecastResponse

Create a value of QueryForecastResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:forecast:QueryForecastResponse', queryForecastResponse_forecast - The forecast.

$sel:httpStatus:QueryForecastResponse', queryForecastResponse_httpStatus - The response's http status code.

Types

DataPoint

data DataPoint

The forecast value for a specific date. Part of the Forecast object.

See: newDataPoint smart constructor.

Constructors

DataPoint' (Maybe Double) (Maybe Text)

newDataPoint :: DataPoint

Create a value of DataPoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:DataPoint', dataPoint_value - The forecast value.

$sel:timestamp:DataPoint', dataPoint_timestamp - The timestamp of the specific forecast.

Forecast

data Forecast

Provides information about a forecast. Returned as part of the QueryForecast response.

See: newForecast smart constructor.

Constructors

Forecast' (Maybe (HashMap Text [DataPoint]))

newForecast :: Forecast

Create a value of Forecast with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:predictions:Forecast', forecast_predictions - The forecast.

The string of the string-to-array map is one of the following values:

p10

p50

p90