Network.AWS.ForecastQuery.Lens

Description

Operations

QueryForecast

queryForecast_endDate :: Lens' QueryForecast (Maybe Text)

The end date for the forecast. Specify the date using this format: yyyy-MM-dd'T'HH:mm:ss (ISO 8601 format). For example, 2015-01-01T20:00:00.

queryForecast_startDate :: Lens' QueryForecast (Maybe Text)

The start date for the forecast. Specify the date using this format: yyyy-MM-dd'T'HH:mm:ss (ISO 8601 format). For example, 2015-01-01T08:00:00.

queryForecast_nextToken :: Lens' QueryForecast (Maybe Text)

If the result of the previous request was truncated, the response includes a NextToken. To retrieve the next set of results, use the token in the next request. Tokens expire after 24 hours.

queryForecast_forecastArn :: Lens' QueryForecast Text

The Amazon Resource Name (ARN) of the forecast to query.

queryForecast_filters :: Lens' QueryForecast (HashMap Text Text)

The filtering criteria to apply when retrieving the forecast. For example, to get the forecast for client_21 in the electricity usage dataset, specify the following:

{"item_id" : "client_21"}

To get the full forecast, use the CreateForecastExportJob operation.

queryForecastResponse_forecast :: Lens' QueryForecastResponse (Maybe Forecast)

The forecast.

queryForecastResponse_httpStatus :: Lens' QueryForecastResponse Int

The response's http status code.

Types

DataPoint

dataPoint_value :: Lens' DataPoint (Maybe Double)

The forecast value.

dataPoint_timestamp :: Lens' DataPoint (Maybe Text)

The timestamp of the specific forecast.

Forecast

forecast_predictions :: Lens' Forecast (Maybe (HashMap Text [DataPoint]))

The forecast.

The string of the string-to-array map is one of the following values:

p10

p50

p90