Network.AWS.ForecastQuery.Types

Description

Service Configuration

defaultService :: Service

API version 2018-06-26 of the Amazon Forecast Query Service SDK configuration.

Errors

_InvalidNextTokenException :: AsError a => Getting (First ServiceError) a ServiceError

The token is not valid. Tokens expire after 24 hours.

_InvalidInputException :: AsError a => Getting (First ServiceError) a ServiceError

The value is invalid or is too long.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

We can't find that resource. Check the information that you've provided and try again.

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The limit on the number of requests per second has been exceeded.

_ResourceInUseException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource is in use.

DataPoint

data DataPoint

The forecast value for a specific date. Part of the Forecast object.

See: newDataPoint smart constructor.

Constructors

DataPoint'

Fields

value :: Maybe Double

The forecast value.

timestamp :: Maybe Text

The timestamp of the specific forecast.

newDataPoint :: DataPoint

Create a value of DataPoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:DataPoint', dataPoint_value - The forecast value.

$sel:timestamp:DataPoint', dataPoint_timestamp - The timestamp of the specific forecast.

dataPoint_value :: Lens' DataPoint (Maybe Double)

The forecast value.

dataPoint_timestamp :: Lens' DataPoint (Maybe Text)

The timestamp of the specific forecast.

Forecast

data Forecast

Provides information about a forecast. Returned as part of the QueryForecast response.

See: newForecast smart constructor.

Constructors

Forecast'

Fields

predictions :: Maybe (HashMap Text [DataPoint])

The forecast.

The string of the string-to-array map is one of the following values:

p10

p50

p90

newForecast :: Forecast

Create a value of Forecast with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:predictions:Forecast', forecast_predictions - The forecast.

The string of the string-to-array map is one of the following values:

p10

p50

p90

forecast_predictions :: Lens' Forecast (Maybe (HashMap Text [DataPoint]))

The forecast.

The string of the string-to-array map is one of the following values:

p10

p50

p90