Network.AWS.FraudDetector.BatchCreateVariable

Description

Creates a batch of variables.

Creating a Request

data BatchCreateVariable

See: newBatchCreateVariable smart constructor.

Constructors

BatchCreateVariable'

Fields

tags :: Maybe [Tag]

A collection of key and value pairs.

variableEntries :: NonEmpty VariableEntry

The list of variables for the batch create variable request.

newBatchCreateVariable

Arguments

:: NonEmpty VariableEntry
$sel:variableEntries:BatchCreateVariable'

-> BatchCreateVariable

Create a value of BatchCreateVariable with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:BatchCreateVariable', batchCreateVariable_tags - A collection of key and value pairs.

$sel:variableEntries:BatchCreateVariable', batchCreateVariable_variableEntries - The list of variables for the batch create variable request.

Request Lenses

batchCreateVariable_tags :: Lens' BatchCreateVariable (Maybe [Tag])

A collection of key and value pairs.

batchCreateVariable_variableEntries :: Lens' BatchCreateVariable (NonEmpty VariableEntry)

The list of variables for the batch create variable request.

Destructuring the Response

data BatchCreateVariableResponse

See: newBatchCreateVariableResponse smart constructor.

Constructors

BatchCreateVariableResponse'

Fields

errors :: Maybe [BatchCreateVariableError]

Provides the errors for the BatchCreateVariable request.

httpStatus :: Int

The response's http status code.

newBatchCreateVariableResponse

Arguments

:: Int
$sel:httpStatus:BatchCreateVariableResponse'

-> BatchCreateVariableResponse

Create a value of BatchCreateVariableResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:errors:BatchCreateVariableResponse', batchCreateVariableResponse_errors - Provides the errors for the BatchCreateVariable request.

$sel:httpStatus:BatchCreateVariableResponse', batchCreateVariableResponse_httpStatus - The response's http status code.

Response Lenses

batchCreateVariableResponse_errors :: Lens' BatchCreateVariableResponse (Maybe [BatchCreateVariableError])

Provides the errors for the BatchCreateVariable request.

batchCreateVariableResponse_httpStatus :: Lens' BatchCreateVariableResponse Int

The response's http status code.