Network.AWS.FraudDetector.CreateBatchImportJob

Description

Creates a batch import job.

Creating a Request

data CreateBatchImportJob

See: newCreateBatchImportJob smart constructor.

Constructors

CreateBatchImportJob'

Fields

tags :: Maybe [Tag]

A collection of key-value pairs associated with this request.

jobId :: Text

The ID of the batch import job. The ID cannot be of a past job, unless the job exists in CREATE_FAILED state.

inputPath :: Text

The URI that points to the Amazon S3 location of your data file.

outputPath :: Text

The URI that points to the Amazon S3 location for storing your results.

eventTypeName :: Text

The name of the event type.

iamRoleArn :: Text

The ARN of the IAM role created for Amazon S3 bucket that holds your data file. The IAM role must have read and write permissions to both input and output S3 buckets.

newCreateBatchImportJob

Arguments

:: Text
$sel:jobId:CreateBatchImportJob'

-> Text
$sel:inputPath:CreateBatchImportJob'

-> Text
$sel:outputPath:CreateBatchImportJob'

-> Text
$sel:eventTypeName:CreateBatchImportJob'

-> Text
$sel:iamRoleArn:CreateBatchImportJob'

-> CreateBatchImportJob

Create a value of CreateBatchImportJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateBatchImportJob', createBatchImportJob_tags - A collection of key-value pairs associated with this request.

$sel:jobId:CreateBatchImportJob', createBatchImportJob_jobId - The ID of the batch import job. The ID cannot be of a past job, unless the job exists in CREATE_FAILED state.

$sel:inputPath:CreateBatchImportJob', createBatchImportJob_inputPath - The URI that points to the Amazon S3 location of your data file.

$sel:outputPath:CreateBatchImportJob', createBatchImportJob_outputPath - The URI that points to the Amazon S3 location for storing your results.

$sel:eventTypeName:CreateBatchImportJob', createBatchImportJob_eventTypeName - The name of the event type.

$sel:iamRoleArn:CreateBatchImportJob', createBatchImportJob_iamRoleArn - The ARN of the IAM role created for Amazon S3 bucket that holds your data file. The IAM role must have read and write permissions to both input and output S3 buckets.

Request Lenses

createBatchImportJob_tags :: Lens' CreateBatchImportJob (Maybe [Tag])

A collection of key-value pairs associated with this request.

createBatchImportJob_jobId :: Lens' CreateBatchImportJob Text

The ID of the batch import job. The ID cannot be of a past job, unless the job exists in CREATE_FAILED state.

createBatchImportJob_inputPath :: Lens' CreateBatchImportJob Text

The URI that points to the Amazon S3 location of your data file.

createBatchImportJob_outputPath :: Lens' CreateBatchImportJob Text

The URI that points to the Amazon S3 location for storing your results.

createBatchImportJob_eventTypeName :: Lens' CreateBatchImportJob Text

The name of the event type.

createBatchImportJob_iamRoleArn :: Lens' CreateBatchImportJob Text

The ARN of the IAM role created for Amazon S3 bucket that holds your data file. The IAM role must have read and write permissions to both input and output S3 buckets.

Destructuring the Response

data CreateBatchImportJobResponse

See: newCreateBatchImportJobResponse smart constructor.

Constructors

CreateBatchImportJobResponse'

Fields

httpStatus :: Int

The response's http status code.

newCreateBatchImportJobResponse

Arguments

:: Int
$sel:httpStatus:CreateBatchImportJobResponse'

-> CreateBatchImportJobResponse

Create a value of CreateBatchImportJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateBatchImportJobResponse', createBatchImportJobResponse_httpStatus - The response's http status code.

Response Lenses

createBatchImportJobResponse_httpStatus :: Lens' CreateBatchImportJobResponse Int

The response's http status code.