Network.AWS.FraudDetector.CreateDetectorVersion

Description

Creates a detector version. The detector version starts in a DRAFT status.

Creating a Request

data CreateDetectorVersion

See: newCreateDetectorVersion smart constructor.

Constructors

CreateDetectorVersion'

Fields

ruleExecutionMode :: Maybe RuleExecutionMode

The rule execution mode for the rules included in the detector version.

You can define and edit the rule mode at the detector version level, when it is in draft status.

If you specify FIRST_MATCHED, Amazon Fraud Detector evaluates rules sequentially, first to last, stopping at the first matched rule. Amazon Fraud dectector then provides the outcomes for that single rule.

If you specifiy ALL_MATCHED, Amazon Fraud Detector evaluates all rules and returns the outcomes for all matched rules.

The default behavior is FIRST_MATCHED.

modelVersions :: Maybe [ModelVersion]

The model versions to include in the detector version.

externalModelEndpoints :: Maybe [Text]

The Amazon Sagemaker model endpoints to include in the detector version.

description :: Maybe Text

The description of the detector version.

tags :: Maybe [Tag]

A collection of key and value pairs.

detectorId :: Text

The ID of the detector under which you want to create a new version.

rules :: [Rule]

The rules to include in the detector version.

newCreateDetectorVersion

Arguments

:: Text
$sel:detectorId:CreateDetectorVersion'

-> CreateDetectorVersion

Create a value of CreateDetectorVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ruleExecutionMode:CreateDetectorVersion', createDetectorVersion_ruleExecutionMode - The rule execution mode for the rules included in the detector version.

You can define and edit the rule mode at the detector version level, when it is in draft status.

If you specify FIRST_MATCHED, Amazon Fraud Detector evaluates rules sequentially, first to last, stopping at the first matched rule. Amazon Fraud dectector then provides the outcomes for that single rule.

If you specifiy ALL_MATCHED, Amazon Fraud Detector evaluates all rules and returns the outcomes for all matched rules.

The default behavior is FIRST_MATCHED.

$sel:modelVersions:CreateDetectorVersion', createDetectorVersion_modelVersions - The model versions to include in the detector version.

$sel:externalModelEndpoints:CreateDetectorVersion', createDetectorVersion_externalModelEndpoints - The Amazon Sagemaker model endpoints to include in the detector version.

$sel:description:CreateDetectorVersion', createDetectorVersion_description - The description of the detector version.

$sel:tags:CreateDetectorVersion', createDetectorVersion_tags - A collection of key and value pairs.

$sel:detectorId:CreateDetectorVersion', createDetectorVersion_detectorId - The ID of the detector under which you want to create a new version.

$sel:rules:CreateDetectorVersion', createDetectorVersion_rules - The rules to include in the detector version.

Request Lenses

createDetectorVersion_ruleExecutionMode :: Lens' CreateDetectorVersion (Maybe RuleExecutionMode)

The rule execution mode for the rules included in the detector version.

You can define and edit the rule mode at the detector version level, when it is in draft status.

If you specify FIRST_MATCHED, Amazon Fraud Detector evaluates rules sequentially, first to last, stopping at the first matched rule. Amazon Fraud dectector then provides the outcomes for that single rule.

If you specifiy ALL_MATCHED, Amazon Fraud Detector evaluates all rules and returns the outcomes for all matched rules.

The default behavior is FIRST_MATCHED.

createDetectorVersion_modelVersions :: Lens' CreateDetectorVersion (Maybe [ModelVersion])

The model versions to include in the detector version.

createDetectorVersion_externalModelEndpoints :: Lens' CreateDetectorVersion (Maybe [Text])

The Amazon Sagemaker model endpoints to include in the detector version.

createDetectorVersion_description :: Lens' CreateDetectorVersion (Maybe Text)

The description of the detector version.

createDetectorVersion_tags :: Lens' CreateDetectorVersion (Maybe [Tag])

A collection of key and value pairs.

createDetectorVersion_detectorId :: Lens' CreateDetectorVersion Text

The ID of the detector under which you want to create a new version.

createDetectorVersion_rules :: Lens' CreateDetectorVersion [Rule]

The rules to include in the detector version.

Destructuring the Response

data CreateDetectorVersionResponse

See: newCreateDetectorVersionResponse smart constructor.

Constructors

CreateDetectorVersionResponse'

Fields

status :: Maybe DetectorVersionStatus

The status of the detector version.

detectorVersionId :: Maybe Text

The ID for the created detector.

detectorId :: Maybe Text

The ID for the created version's parent detector.

httpStatus :: Int

The response's http status code.

newCreateDetectorVersionResponse

Arguments

:: Int
$sel:httpStatus:CreateDetectorVersionResponse'

-> CreateDetectorVersionResponse

Create a value of CreateDetectorVersionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:CreateDetectorVersionResponse', createDetectorVersionResponse_status - The status of the detector version.

$sel:detectorVersionId:CreateDetectorVersionResponse', createDetectorVersionResponse_detectorVersionId - The ID for the created detector.

$sel:detectorId:CreateDetectorVersion', createDetectorVersionResponse_detectorId - The ID for the created version's parent detector.

$sel:httpStatus:CreateDetectorVersionResponse', createDetectorVersionResponse_httpStatus - The response's http status code.

Response Lenses

createDetectorVersionResponse_status :: Lens' CreateDetectorVersionResponse (Maybe DetectorVersionStatus)

The status of the detector version.

createDetectorVersionResponse_detectorVersionId :: Lens' CreateDetectorVersionResponse (Maybe Text)

The ID for the created detector.

createDetectorVersionResponse_detectorId :: Lens' CreateDetectorVersionResponse (Maybe Text)

The ID for the created version's parent detector.

createDetectorVersionResponse_httpStatus :: Lens' CreateDetectorVersionResponse Int

The response's http status code.