Network.AWS.FraudDetector.DeleteDetectorVersion

Description

Deletes the detector version. You cannot delete detector versions that are in ACTIVE status.

When you delete a detector version, Amazon Fraud Detector permanently deletes the detector and the data is no longer stored in Amazon Fraud Detector.

Creating a Request

data DeleteDetectorVersion

See: newDeleteDetectorVersion smart constructor.

Constructors

DeleteDetectorVersion'

Fields

detectorId :: Text

The ID of the parent detector for the detector version to delete.

detectorVersionId :: Text

The ID of the detector version to delete.

newDeleteDetectorVersion

Arguments

:: Text
$sel:detectorId:DeleteDetectorVersion'

-> Text
$sel:detectorVersionId:DeleteDetectorVersion'

-> DeleteDetectorVersion

Create a value of DeleteDetectorVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:detectorId:DeleteDetectorVersion', deleteDetectorVersion_detectorId - The ID of the parent detector for the detector version to delete.

$sel:detectorVersionId:DeleteDetectorVersion', deleteDetectorVersion_detectorVersionId - The ID of the detector version to delete.

Request Lenses

deleteDetectorVersion_detectorId :: Lens' DeleteDetectorVersion Text

The ID of the parent detector for the detector version to delete.

deleteDetectorVersion_detectorVersionId :: Lens' DeleteDetectorVersion Text

The ID of the detector version to delete.

Destructuring the Response

data DeleteDetectorVersionResponse

See: newDeleteDetectorVersionResponse smart constructor.

Constructors

DeleteDetectorVersionResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteDetectorVersionResponse

Arguments

:: Int
$sel:httpStatus:DeleteDetectorVersionResponse'

-> DeleteDetectorVersionResponse

Create a value of DeleteDetectorVersionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteDetectorVersionResponse', deleteDetectorVersionResponse_httpStatus - The response's http status code.

Response Lenses

deleteDetectorVersionResponse_httpStatus :: Lens' DeleteDetectorVersionResponse Int

The response's http status code.