Network.AWS.FraudDetector.DeleteExternalModel

Description

Removes a SageMaker model from Amazon Fraud Detector.

You can remove an Amazon SageMaker model if it is not associated with a detector version. Removing a SageMaker model disconnects it from Amazon Fraud Detector, but the model remains available in SageMaker.

Creating a Request

data DeleteExternalModel

See: newDeleteExternalModel smart constructor.

Constructors

DeleteExternalModel'

Fields

modelEndpoint :: Text

The endpoint of the Amazon Sagemaker model to delete.

newDeleteExternalModel

Arguments

:: Text
$sel:modelEndpoint:DeleteExternalModel'

-> DeleteExternalModel

Create a value of DeleteExternalModel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:modelEndpoint:DeleteExternalModel', deleteExternalModel_modelEndpoint - The endpoint of the Amazon Sagemaker model to delete.

Request Lenses

deleteExternalModel_modelEndpoint :: Lens' DeleteExternalModel Text

The endpoint of the Amazon Sagemaker model to delete.

Destructuring the Response

data DeleteExternalModelResponse

See: newDeleteExternalModelResponse smart constructor.

Constructors

DeleteExternalModelResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteExternalModelResponse

Arguments

:: Int
$sel:httpStatus:DeleteExternalModelResponse'

-> DeleteExternalModelResponse

Create a value of DeleteExternalModelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteExternalModelResponse', deleteExternalModelResponse_httpStatus - The response's http status code.

Response Lenses

deleteExternalModelResponse_httpStatus :: Lens' DeleteExternalModelResponse Int

The response's http status code.