Network.AWS.FraudDetector.DeleteModelVersion

Description

Deletes a model version.

You can delete models and model versions in Amazon Fraud Detector, provided that they are not associated with a detector version.

When you delete a model version, Amazon Fraud Detector permanently deletes that model version and the data is no longer stored in Amazon Fraud Detector.

Creating a Request

data DeleteModelVersion

See: newDeleteModelVersion smart constructor.

Constructors

DeleteModelVersion'

Fields

modelId :: Text

The model ID of the model version to delete.

modelType :: ModelTypeEnum

The model type of the model version to delete.

modelVersionNumber :: Text

The model version number of the model version to delete.

newDeleteModelVersion

Arguments

:: Text
$sel:modelId:DeleteModelVersion'

-> ModelTypeEnum
$sel:modelType:DeleteModelVersion'

-> Text
$sel:modelVersionNumber:DeleteModelVersion'

-> DeleteModelVersion

Create a value of DeleteModelVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:modelId:DeleteModelVersion', deleteModelVersion_modelId - The model ID of the model version to delete.

$sel:modelType:DeleteModelVersion', deleteModelVersion_modelType - The model type of the model version to delete.

$sel:modelVersionNumber:DeleteModelVersion', deleteModelVersion_modelVersionNumber - The model version number of the model version to delete.

Request Lenses

deleteModelVersion_modelId :: Lens' DeleteModelVersion Text

The model ID of the model version to delete.

deleteModelVersion_modelType :: Lens' DeleteModelVersion ModelTypeEnum

The model type of the model version to delete.

deleteModelVersion_modelVersionNumber :: Lens' DeleteModelVersion Text

The model version number of the model version to delete.

Destructuring the Response

data DeleteModelVersionResponse

See: newDeleteModelVersionResponse smart constructor.

Constructors

DeleteModelVersionResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteModelVersionResponse

Arguments

:: Int
$sel:httpStatus:DeleteModelVersionResponse'

-> DeleteModelVersionResponse

Create a value of DeleteModelVersionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteModelVersionResponse', deleteModelVersionResponse_httpStatus - The response's http status code.

Response Lenses

deleteModelVersionResponse_httpStatus :: Lens' DeleteModelVersionResponse Int

The response's http status code.