Network.AWS.FraudDetector.GetDetectorVersion

Description

Gets a particular detector version.

Creating a Request

data GetDetectorVersion

See: newGetDetectorVersion smart constructor.

Constructors

GetDetectorVersion'

Fields

detectorId :: Text

The detector ID.

detectorVersionId :: Text

The detector version ID.

newGetDetectorVersion

Arguments

:: Text
$sel:detectorId:GetDetectorVersion'

-> Text
$sel:detectorVersionId:GetDetectorVersion'

-> GetDetectorVersion

Create a value of GetDetectorVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:detectorId:GetDetectorVersion', getDetectorVersion_detectorId - The detector ID.

$sel:detectorVersionId:GetDetectorVersion', getDetectorVersion_detectorVersionId - The detector version ID.

Request Lenses

getDetectorVersion_detectorId :: Lens' GetDetectorVersion Text

The detector ID.

getDetectorVersion_detectorVersionId :: Lens' GetDetectorVersion Text

The detector version ID.

Destructuring the Response

data GetDetectorVersionResponse

See: newGetDetectorVersionResponse smart constructor.

Constructors

GetDetectorVersionResponse'

Fields

status :: Maybe DetectorVersionStatus

The status of the detector version.

ruleExecutionMode :: Maybe RuleExecutionMode

The execution mode of the rule in the dectector

FIRST_MATCHED indicates that Amazon Fraud Detector evaluates rules sequentially, first to last, stopping at the first matched rule. Amazon Fraud dectector then provides the outcomes for that single rule.

ALL_MATCHED indicates that Amazon Fraud Detector evaluates all rules and returns the outcomes for all matched rules. You can define and edit the rule mode at the detector version level, when it is in draft status.

lastUpdatedTime :: Maybe Text

The timestamp when the detector version was last updated.

arn :: Maybe Text

The detector version ARN.

rules :: Maybe [Rule]

The rules included in the detector version.

detectorVersionId :: Maybe Text

The detector version ID.

createdTime :: Maybe Text

The timestamp when the detector version was created.

modelVersions :: Maybe [ModelVersion]

The model versions included in the detector version.

detectorId :: Maybe Text

The detector ID.

externalModelEndpoints :: Maybe [Text]

The Amazon SageMaker model endpoints included in the detector version.

description :: Maybe Text

The detector version description.

httpStatus :: Int

The response's http status code.

newGetDetectorVersionResponse

Arguments

:: Int
$sel:httpStatus:GetDetectorVersionResponse'

-> GetDetectorVersionResponse

Create a value of GetDetectorVersionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:GetDetectorVersionResponse', getDetectorVersionResponse_status - The status of the detector version.

$sel:ruleExecutionMode:GetDetectorVersionResponse', getDetectorVersionResponse_ruleExecutionMode - The execution mode of the rule in the dectector

FIRST_MATCHED indicates that Amazon Fraud Detector evaluates rules sequentially, first to last, stopping at the first matched rule. Amazon Fraud dectector then provides the outcomes for that single rule.

ALL_MATCHED indicates that Amazon Fraud Detector evaluates all rules and returns the outcomes for all matched rules. You can define and edit the rule mode at the detector version level, when it is in draft status.

$sel:lastUpdatedTime:GetDetectorVersionResponse', getDetectorVersionResponse_lastUpdatedTime - The timestamp when the detector version was last updated.

$sel:arn:GetDetectorVersionResponse', getDetectorVersionResponse_arn - The detector version ARN.

$sel:rules:GetDetectorVersionResponse', getDetectorVersionResponse_rules - The rules included in the detector version.

$sel:detectorVersionId:GetDetectorVersion', getDetectorVersionResponse_detectorVersionId - The detector version ID.

$sel:createdTime:GetDetectorVersionResponse', getDetectorVersionResponse_createdTime - The timestamp when the detector version was created.

$sel:modelVersions:GetDetectorVersionResponse', getDetectorVersionResponse_modelVersions - The model versions included in the detector version.

$sel:detectorId:GetDetectorVersion', getDetectorVersionResponse_detectorId - The detector ID.

$sel:externalModelEndpoints:GetDetectorVersionResponse', getDetectorVersionResponse_externalModelEndpoints - The Amazon SageMaker model endpoints included in the detector version.

$sel:description:GetDetectorVersionResponse', getDetectorVersionResponse_description - The detector version description.

$sel:httpStatus:GetDetectorVersionResponse', getDetectorVersionResponse_httpStatus - The response's http status code.

Response Lenses

getDetectorVersionResponse_status :: Lens' GetDetectorVersionResponse (Maybe DetectorVersionStatus)

The status of the detector version.

getDetectorVersionResponse_ruleExecutionMode :: Lens' GetDetectorVersionResponse (Maybe RuleExecutionMode)

The execution mode of the rule in the dectector

FIRST_MATCHED indicates that Amazon Fraud Detector evaluates rules sequentially, first to last, stopping at the first matched rule. Amazon Fraud dectector then provides the outcomes for that single rule.

ALL_MATCHED indicates that Amazon Fraud Detector evaluates all rules and returns the outcomes for all matched rules. You can define and edit the rule mode at the detector version level, when it is in draft status.

getDetectorVersionResponse_lastUpdatedTime :: Lens' GetDetectorVersionResponse (Maybe Text)

The timestamp when the detector version was last updated.

getDetectorVersionResponse_arn :: Lens' GetDetectorVersionResponse (Maybe Text)

The detector version ARN.

getDetectorVersionResponse_rules :: Lens' GetDetectorVersionResponse (Maybe [Rule])

The rules included in the detector version.

getDetectorVersionResponse_detectorVersionId :: Lens' GetDetectorVersionResponse (Maybe Text)

The detector version ID.

getDetectorVersionResponse_createdTime :: Lens' GetDetectorVersionResponse (Maybe Text)

The timestamp when the detector version was created.

getDetectorVersionResponse_modelVersions :: Lens' GetDetectorVersionResponse (Maybe [ModelVersion])

The model versions included in the detector version.

getDetectorVersionResponse_detectorId :: Lens' GetDetectorVersionResponse (Maybe Text)

The detector ID.

getDetectorVersionResponse_externalModelEndpoints :: Lens' GetDetectorVersionResponse (Maybe [Text])

The Amazon SageMaker model endpoints included in the detector version.

getDetectorVersionResponse_description :: Lens' GetDetectorVersionResponse (Maybe Text)

The detector version description.

getDetectorVersionResponse_httpStatus :: Lens' GetDetectorVersionResponse Int

The response's http status code.