Network.AWS.FraudDetector.GetEventPrediction

Description

Evaluates an event against a detector version. If a version ID is not provided, the detector’s (ACTIVE) version is used.

Creating a Request

data GetEventPrediction

See: newGetEventPrediction smart constructor.

Constructors

GetEventPrediction'

Fields

detectorVersionId :: Maybe Text

The detector version ID.

externalModelEndpointDataBlobs :: Maybe (Sensitive (HashMap Text ModelEndpointDataBlob))

The Amazon SageMaker model endpoint input data blobs.

detectorId :: Text

The detector ID.

eventId :: Text

The unique ID used to identify the event.

eventTypeName :: Text

The event type associated with the detector specified for the prediction.

entities :: [Sensitive Entity]

The entity type (associated with the detector's event type) and specific entity ID representing who performed the event. If an entity id is not available, use "UNKNOWN."

eventTimestamp :: Text

Timestamp that defines when the event under evaluation occurred. The timestamp must be specified using ISO 8601 standard in UTC.

eventVariables :: HashMap Text (Sensitive Text)

Names of the event type's variables you defined in Amazon Fraud Detector to represent data elements and their corresponding values for the event you are sending for evaluation.

You must provide at least one eventVariable

To ensure most accurate fraud prediction and to simplify your data preparation, Amazon Fraud Detector will replace all missing variables or values as follows:

For Amazon Fraud Detector trained models:

If a null value is provided explicitly for a variable or if a variable is missing, model will replace the null value or the missing variable (no variable name in the eventVariables map) with calculated default mean/medians for numeric variables and with special values for categorical variables.

For imported SageMaker models:

If a null value is provided explicitly for a variable, the model and rules will use “null” as the value. If a variable is not provided (no variable name in the eventVariables map), model and rules will use the default value that is provided for the variable.

newGetEventPrediction

Arguments

:: Text
$sel:detectorId:GetEventPrediction'

-> Text
$sel:eventId:GetEventPrediction'

-> Text
$sel:eventTypeName:GetEventPrediction'

-> Text
$sel:eventTimestamp:GetEventPrediction'

-> GetEventPrediction

Create a value of GetEventPrediction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:detectorVersionId:GetEventPrediction', getEventPrediction_detectorVersionId - The detector version ID.

$sel:externalModelEndpointDataBlobs:GetEventPrediction', getEventPrediction_externalModelEndpointDataBlobs - The Amazon SageMaker model endpoint input data blobs.

$sel:detectorId:GetEventPrediction', getEventPrediction_detectorId - The detector ID.

$sel:eventId:GetEventPrediction', getEventPrediction_eventId - The unique ID used to identify the event.

$sel:eventTypeName:GetEventPrediction', getEventPrediction_eventTypeName - The event type associated with the detector specified for the prediction.

$sel:entities:GetEventPrediction', getEventPrediction_entities - The entity type (associated with the detector's event type) and specific entity ID representing who performed the event. If an entity id is not available, use "UNKNOWN."

$sel:eventTimestamp:GetEventPrediction', getEventPrediction_eventTimestamp - Timestamp that defines when the event under evaluation occurred. The timestamp must be specified using ISO 8601 standard in UTC.

$sel:eventVariables:GetEventPrediction', getEventPrediction_eventVariables - Names of the event type's variables you defined in Amazon Fraud Detector to represent data elements and their corresponding values for the event you are sending for evaluation.

You must provide at least one eventVariable

To ensure most accurate fraud prediction and to simplify your data preparation, Amazon Fraud Detector will replace all missing variables or values as follows:

For Amazon Fraud Detector trained models:

If a null value is provided explicitly for a variable or if a variable is missing, model will replace the null value or the missing variable (no variable name in the eventVariables map) with calculated default mean/medians for numeric variables and with special values for categorical variables.

For imported SageMaker models:

If a null value is provided explicitly for a variable, the model and rules will use “null” as the value. If a variable is not provided (no variable name in the eventVariables map), model and rules will use the default value that is provided for the variable.

Request Lenses

getEventPrediction_detectorVersionId :: Lens' GetEventPrediction (Maybe Text)

The detector version ID.

getEventPrediction_externalModelEndpointDataBlobs :: Lens' GetEventPrediction (Maybe (HashMap Text ModelEndpointDataBlob))

The Amazon SageMaker model endpoint input data blobs.

getEventPrediction_detectorId :: Lens' GetEventPrediction Text

The detector ID.

getEventPrediction_eventId :: Lens' GetEventPrediction Text

The unique ID used to identify the event.

getEventPrediction_eventTypeName :: Lens' GetEventPrediction Text

The event type associated with the detector specified for the prediction.

getEventPrediction_entities :: Lens' GetEventPrediction [Entity]

The entity type (associated with the detector's event type) and specific entity ID representing who performed the event. If an entity id is not available, use "UNKNOWN."

getEventPrediction_eventTimestamp :: Lens' GetEventPrediction Text

Timestamp that defines when the event under evaluation occurred. The timestamp must be specified using ISO 8601 standard in UTC.

getEventPrediction_eventVariables :: Lens' GetEventPrediction (HashMap Text Text)

Names of the event type's variables you defined in Amazon Fraud Detector to represent data elements and their corresponding values for the event you are sending for evaluation.

You must provide at least one eventVariable

To ensure most accurate fraud prediction and to simplify your data preparation, Amazon Fraud Detector will replace all missing variables or values as follows:

For Amazon Fraud Detector trained models:

If a null value is provided explicitly for a variable or if a variable is missing, model will replace the null value or the missing variable (no variable name in the eventVariables map) with calculated default mean/medians for numeric variables and with special values for categorical variables.

For imported SageMaker models:

If a null value is provided explicitly for a variable, the model and rules will use “null” as the value. If a variable is not provided (no variable name in the eventVariables map), model and rules will use the default value that is provided for the variable.

Destructuring the Response

data GetEventPredictionResponse

See: newGetEventPredictionResponse smart constructor.

Constructors

GetEventPredictionResponse'

Fields

modelScores :: Maybe [ModelScores]

The model scores. Amazon Fraud Detector generates model scores between 0 and 1000, where 0 is low fraud risk and 1000 is high fraud risk. Model scores are directly related to the false positive rate (FPR). For example, a score of 600 corresponds to an estimated 10% false positive rate whereas a score of 900 corresponds to an estimated 2% false positive rate.

externalModelOutputs :: Maybe [ExternalModelOutputs]

The model scores for Amazon SageMaker models.

ruleResults :: Maybe [RuleResult]

The results from the rules.

httpStatus :: Int

The response's http status code.

newGetEventPredictionResponse

Arguments

:: Int
$sel:httpStatus:GetEventPredictionResponse'

-> GetEventPredictionResponse

Create a value of GetEventPredictionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:modelScores:GetEventPredictionResponse', getEventPredictionResponse_modelScores - The model scores. Amazon Fraud Detector generates model scores between 0 and 1000, where 0 is low fraud risk and 1000 is high fraud risk. Model scores are directly related to the false positive rate (FPR). For example, a score of 600 corresponds to an estimated 10% false positive rate whereas a score of 900 corresponds to an estimated 2% false positive rate.

$sel:externalModelOutputs:GetEventPredictionResponse', getEventPredictionResponse_externalModelOutputs - The model scores for Amazon SageMaker models.

$sel:ruleResults:GetEventPredictionResponse', getEventPredictionResponse_ruleResults - The results from the rules.

$sel:httpStatus:GetEventPredictionResponse', getEventPredictionResponse_httpStatus - The response's http status code.

Response Lenses

getEventPredictionResponse_modelScores :: Lens' GetEventPredictionResponse (Maybe [ModelScores])

The model scores. Amazon Fraud Detector generates model scores between 0 and 1000, where 0 is low fraud risk and 1000 is high fraud risk. Model scores are directly related to the false positive rate (FPR). For example, a score of 600 corresponds to an estimated 10% false positive rate whereas a score of 900 corresponds to an estimated 2% false positive rate.

getEventPredictionResponse_externalModelOutputs :: Lens' GetEventPredictionResponse (Maybe [ExternalModelOutputs])

The model scores for Amazon SageMaker models.

getEventPredictionResponse_ruleResults :: Lens' GetEventPredictionResponse (Maybe [RuleResult])

The results from the rules.

getEventPredictionResponse_httpStatus :: Lens' GetEventPredictionResponse Int

The response's http status code.