Network.AWS.FraudDetector.GetLabels

Description

Gets all labels or a specific label if name is provided. This is a paginated API. If you provide a null maxResults, this action retrieves a maximum of 50 records per page. If you provide a maxResults, the value must be between 10 and 50. To get the next page results, provide the pagination token from the GetGetLabelsResponse as part of your request. A null pagination token fetches the records from the beginning.

Creating a Request

data GetLabels

See: newGetLabels smart constructor.

Constructors

GetLabels'

Fields

nextToken :: Maybe Text

The next token for the subsequent request.

name :: Maybe Text

The name of the label or labels to get.

maxResults :: Maybe Natural

The maximum number of objects to return for the request.

newGetLabels :: GetLabels

Create a value of GetLabels with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetLabels', getLabels_nextToken - The next token for the subsequent request.

$sel:name:GetLabels', getLabels_name - The name of the label or labels to get.

$sel:maxResults:GetLabels', getLabels_maxResults - The maximum number of objects to return for the request.

Request Lenses

getLabels_nextToken :: Lens' GetLabels (Maybe Text)

The next token for the subsequent request.

getLabels_name :: Lens' GetLabels (Maybe Text)

The name of the label or labels to get.

getLabels_maxResults :: Lens' GetLabels (Maybe Natural)

The maximum number of objects to return for the request.

Destructuring the Response

data GetLabelsResponse

See: newGetLabelsResponse smart constructor.

Constructors

GetLabelsResponse'

Fields

nextToken :: Maybe Text

The next page token.

labels :: Maybe [Label]

An array of labels.

httpStatus :: Int

The response's http status code.

newGetLabelsResponse

Arguments

:: Int
$sel:httpStatus:GetLabelsResponse'

-> GetLabelsResponse

Create a value of GetLabelsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetLabels', getLabelsResponse_nextToken - The next page token.

$sel:labels:GetLabelsResponse', getLabelsResponse_labels - An array of labels.

$sel:httpStatus:GetLabelsResponse', getLabelsResponse_httpStatus - The response's http status code.

Response Lenses

getLabelsResponse_nextToken :: Lens' GetLabelsResponse (Maybe Text)

The next page token.

getLabelsResponse_labels :: Lens' GetLabelsResponse (Maybe [Label])

An array of labels.

getLabelsResponse_httpStatus :: Lens' GetLabelsResponse Int

The response's http status code.