Network.AWS.FraudDetector.GetModels

Description

Gets one or more models. Gets all models for the Amazon Web Services account if no model type and no model id provided. Gets all models for the Amazon Web Services account and model type, if the model type is specified but model id is not provided. Gets a specific model if (model type, model id) tuple is specified.

This is a paginated API. If you provide a null maxResults, this action retrieves a maximum of 10 records per page. If you provide a maxResults, the value must be between 1 and 10. To get the next page results, provide the pagination token from the response as part of your request. A null pagination token fetches the records from the beginning.

Creating a Request

data GetModels

See: newGetModels smart constructor.

Constructors

GetModels'

Fields

modelType :: Maybe ModelTypeEnum

The model type.

modelId :: Maybe Text

The model ID.

nextToken :: Maybe Text

The next token for the subsequent request.

maxResults :: Maybe Natural

The maximum number of objects to return for the request.

newGetModels :: GetModels

Create a value of GetModels with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:modelType:GetModels', getModels_modelType - The model type.

$sel:modelId:GetModels', getModels_modelId - The model ID.

$sel:nextToken:GetModels', getModels_nextToken - The next token for the subsequent request.

$sel:maxResults:GetModels', getModels_maxResults - The maximum number of objects to return for the request.

Request Lenses

getModels_modelType :: Lens' GetModels (Maybe ModelTypeEnum)

The model type.

getModels_modelId :: Lens' GetModels (Maybe Text)

The model ID.

getModels_nextToken :: Lens' GetModels (Maybe Text)

The next token for the subsequent request.

getModels_maxResults :: Lens' GetModels (Maybe Natural)

The maximum number of objects to return for the request.

Destructuring the Response

data GetModelsResponse

See: newGetModelsResponse smart constructor.

Constructors

GetModelsResponse'

Fields

models :: Maybe [Model]

The array of models.

nextToken :: Maybe Text

The next page token to be used in subsequent requests.

httpStatus :: Int

The response's http status code.

newGetModelsResponse

Arguments

:: Int
$sel:httpStatus:GetModelsResponse'

-> GetModelsResponse

Create a value of GetModelsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:models:GetModelsResponse', getModelsResponse_models - The array of models.

$sel:nextToken:GetModels', getModelsResponse_nextToken - The next page token to be used in subsequent requests.

$sel:httpStatus:GetModelsResponse', getModelsResponse_httpStatus - The response's http status code.

Response Lenses

getModelsResponse_models :: Lens' GetModelsResponse (Maybe [Model])

The array of models.

getModelsResponse_nextToken :: Lens' GetModelsResponse (Maybe Text)

The next page token to be used in subsequent requests.

getModelsResponse_httpStatus :: Lens' GetModelsResponse Int

The response's http status code.