Network.AWS.FraudDetector.GetRules

Description

Get all rules for a detector (paginated) if ruleId and ruleVersion are not specified. Gets all rules for the detector and the ruleId if present (paginated). Gets a specific rule if both the ruleId and the ruleVersion are specified.

This is a paginated API. Providing null maxResults results in retrieving maximum of 100 records per page. If you provide maxResults the value must be between 50 and 100. To get the next page result, a provide a pagination token from GetRulesResult as part of your request. Null pagination token fetches the records from the beginning.

Creating a Request

data GetRules

See: newGetRules smart constructor.

Constructors

GetRules'

Fields

ruleVersion :: Maybe Text

The rule version.

ruleId :: Maybe Text

The rule ID.

nextToken :: Maybe Text

The next page token.

maxResults :: Maybe Natural

The maximum number of rules to return for the request.

detectorId :: Text

The detector ID.

newGetRules

Arguments

:: Text
$sel:detectorId:GetRules'

-> GetRules

Create a value of GetRules with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ruleVersion:GetRules', getRules_ruleVersion - The rule version.

$sel:ruleId:GetRules', getRules_ruleId - The rule ID.

$sel:nextToken:GetRules', getRules_nextToken - The next page token.

$sel:maxResults:GetRules', getRules_maxResults - The maximum number of rules to return for the request.

$sel:detectorId:GetRules', getRules_detectorId - The detector ID.

Request Lenses

getRules_ruleVersion :: Lens' GetRules (Maybe Text)

The rule version.

getRules_ruleId :: Lens' GetRules (Maybe Text)

The rule ID.

getRules_nextToken :: Lens' GetRules (Maybe Text)

The next page token.

getRules_maxResults :: Lens' GetRules (Maybe Natural)

The maximum number of rules to return for the request.

getRules_detectorId :: Lens' GetRules Text

The detector ID.

Destructuring the Response

data GetRulesResponse

See: newGetRulesResponse smart constructor.

Constructors

GetRulesResponse'

Fields

ruleDetails :: Maybe [RuleDetail]

The details of the requested rule.

nextToken :: Maybe Text

The next page token to be used in subsequent requests.

httpStatus :: Int

The response's http status code.

newGetRulesResponse

Arguments

:: Int
$sel:httpStatus:GetRulesResponse'

-> GetRulesResponse

Create a value of GetRulesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ruleDetails:GetRulesResponse', getRulesResponse_ruleDetails - The details of the requested rule.

$sel:nextToken:GetRules', getRulesResponse_nextToken - The next page token to be used in subsequent requests.

$sel:httpStatus:GetRulesResponse', getRulesResponse_httpStatus - The response's http status code.

Response Lenses

getRulesResponse_ruleDetails :: Lens' GetRulesResponse (Maybe [RuleDetail])

The details of the requested rule.

getRulesResponse_nextToken :: Lens' GetRulesResponse (Maybe Text)

The next page token to be used in subsequent requests.

getRulesResponse_httpStatus :: Lens' GetRulesResponse Int

The response's http status code.