Network.AWS.FraudDetector.PutEventType

Description

Creates or updates an event type. An event is a business activity that is evaluated for fraud risk. With Amazon Fraud Detector, you generate fraud predictions for events. An event type defines the structure for an event sent to Amazon Fraud Detector. This includes the variables sent as part of the event, the entity performing the event (such as a customer), and the labels that classify the event. Example event types include online payment transactions, account registrations, and authentications.

Creating a Request

data PutEventType

See: newPutEventType smart constructor.

Constructors

PutEventType'

Fields

labels :: Maybe [Text]

The event type labels.

eventIngestion :: Maybe EventIngestion

Specifies if ingenstion is enabled or disabled.

description :: Maybe Text

The description of the event type.

tags :: Maybe [Tag]

A collection of key and value pairs.

name :: Text

The name.

eventVariables :: NonEmpty Text

The event type variables.

entityTypes :: NonEmpty Text

The entity type for the event type. Example entity types: customer, merchant, account.

newPutEventType

Arguments

:: Text
$sel:name:PutEventType'

-> NonEmpty Text
$sel:eventVariables:PutEventType'

-> NonEmpty Text
$sel:entityTypes:PutEventType'

-> PutEventType

Create a value of PutEventType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:labels:PutEventType', putEventType_labels - The event type labels.

$sel:eventIngestion:PutEventType', putEventType_eventIngestion - Specifies if ingenstion is enabled or disabled.

$sel:description:PutEventType', putEventType_description - The description of the event type.

$sel:tags:PutEventType', putEventType_tags - A collection of key and value pairs.

$sel:name:PutEventType', putEventType_name - The name.

$sel:eventVariables:PutEventType', putEventType_eventVariables - The event type variables.

$sel:entityTypes:PutEventType', putEventType_entityTypes - The entity type for the event type. Example entity types: customer, merchant, account.

Request Lenses

putEventType_labels :: Lens' PutEventType (Maybe [Text])

The event type labels.

putEventType_eventIngestion :: Lens' PutEventType (Maybe EventIngestion)

Specifies if ingenstion is enabled or disabled.

putEventType_description :: Lens' PutEventType (Maybe Text)

The description of the event type.

putEventType_tags :: Lens' PutEventType (Maybe [Tag])

A collection of key and value pairs.

putEventType_name :: Lens' PutEventType Text

The name.

putEventType_eventVariables :: Lens' PutEventType (NonEmpty Text)

The event type variables.

putEventType_entityTypes :: Lens' PutEventType (NonEmpty Text)

The entity type for the event type. Example entity types: customer, merchant, account.

Destructuring the Response

data PutEventTypeResponse

See: newPutEventTypeResponse smart constructor.

Constructors

PutEventTypeResponse'

Fields

httpStatus :: Int

The response's http status code.

newPutEventTypeResponse

Arguments

:: Int
$sel:httpStatus:PutEventTypeResponse'

-> PutEventTypeResponse

Create a value of PutEventTypeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:PutEventTypeResponse', putEventTypeResponse_httpStatus - The response's http status code.

Response Lenses

putEventTypeResponse_httpStatus :: Lens' PutEventTypeResponse Int

The response's http status code.