Network.AWS.FraudDetector.PutExternalModel

Description

Creates or updates an Amazon SageMaker model endpoint. You can also use this action to update the configuration of the model endpoint, including the IAM role and/or the mapped variables.

Creating a Request

data PutExternalModel

See: newPutExternalModel smart constructor.

Constructors

PutExternalModel'

Fields

tags :: Maybe [Tag]

A collection of key and value pairs.

modelEndpoint :: Text

The model endpoints name.

modelSource :: ModelSource

The source of the model.

invokeModelEndpointRoleArn :: Text

The IAM role used to invoke the model endpoint.

inputConfiguration :: ModelInputConfiguration

The model endpoint input configuration.

outputConfiguration :: ModelOutputConfiguration

The model endpoint output configuration.

modelEndpointStatus :: ModelEndpointStatus

The model endpoint’s status in Amazon Fraud Detector.

newPutExternalModel

Arguments

:: Text
$sel:modelEndpoint:PutExternalModel'

-> ModelSource
$sel:modelSource:PutExternalModel'

-> Text
$sel:invokeModelEndpointRoleArn:PutExternalModel'

-> ModelInputConfiguration
$sel:inputConfiguration:PutExternalModel'

-> ModelOutputConfiguration
$sel:outputConfiguration:PutExternalModel'

-> ModelEndpointStatus
$sel:modelEndpointStatus:PutExternalModel'

-> PutExternalModel

Create a value of PutExternalModel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:PutExternalModel', putExternalModel_tags - A collection of key and value pairs.

$sel:modelEndpoint:PutExternalModel', putExternalModel_modelEndpoint - The model endpoints name.

$sel:modelSource:PutExternalModel', putExternalModel_modelSource - The source of the model.

$sel:invokeModelEndpointRoleArn:PutExternalModel', putExternalModel_invokeModelEndpointRoleArn - The IAM role used to invoke the model endpoint.

$sel:inputConfiguration:PutExternalModel', putExternalModel_inputConfiguration - The model endpoint input configuration.

$sel:outputConfiguration:PutExternalModel', putExternalModel_outputConfiguration - The model endpoint output configuration.

$sel:modelEndpointStatus:PutExternalModel', putExternalModel_modelEndpointStatus - The model endpoint’s status in Amazon Fraud Detector.

Request Lenses

putExternalModel_tags :: Lens' PutExternalModel (Maybe [Tag])

A collection of key and value pairs.

putExternalModel_modelEndpoint :: Lens' PutExternalModel Text

The model endpoints name.

putExternalModel_modelSource :: Lens' PutExternalModel ModelSource

The source of the model.

putExternalModel_invokeModelEndpointRoleArn :: Lens' PutExternalModel Text

The IAM role used to invoke the model endpoint.

putExternalModel_inputConfiguration :: Lens' PutExternalModel ModelInputConfiguration

The model endpoint input configuration.

putExternalModel_outputConfiguration :: Lens' PutExternalModel ModelOutputConfiguration

The model endpoint output configuration.

putExternalModel_modelEndpointStatus :: Lens' PutExternalModel ModelEndpointStatus

The model endpoint’s status in Amazon Fraud Detector.

Destructuring the Response

data PutExternalModelResponse

See: newPutExternalModelResponse smart constructor.

Constructors

PutExternalModelResponse'

Fields

httpStatus :: Int

The response's http status code.

newPutExternalModelResponse

Arguments

:: Int
$sel:httpStatus:PutExternalModelResponse'

-> PutExternalModelResponse

Create a value of PutExternalModelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:PutExternalModelResponse', putExternalModelResponse_httpStatus - The response's http status code.

Response Lenses

putExternalModelResponse_httpStatus :: Lens' PutExternalModelResponse Int

The response's http status code.