Network.AWS.FraudDetector.Types.BatchImport

Description

Documentation

data BatchImport

The batch import job details.

See: newBatchImport smart constructor.

Constructors

BatchImport'

Fields

failureReason :: Maybe Text

The reason batch import job failed.

iamRoleArn :: Maybe Text

The ARN of the IAM role to use for this job request.

status :: Maybe AsyncJobStatus

The status of the batch import job.

processedRecordsCount :: Maybe Int

The number of records processed by batch import job.

totalRecordsCount :: Maybe Int

The total number of records in the batch import job.

jobId :: Maybe Text

The ID of the batch import job.

arn :: Maybe Text

The ARN of the batch import job.

startTime :: Maybe Text

Timestamp of when the batch import job started.

eventTypeName :: Maybe Text

The name of the event type.

completionTime :: Maybe Text

Timestamp of when batch import job completed.

outputPath :: Maybe Text

The Amazon S3 location of your output file.

inputPath :: Maybe Text

The Amazon S3 location of your data file for batch import.

failedRecordsCount :: Maybe Int

The number of records that failed to import.

newBatchImport :: BatchImport

Create a value of BatchImport with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:BatchImport', batchImport_failureReason - The reason batch import job failed.

$sel:iamRoleArn:BatchImport', batchImport_iamRoleArn - The ARN of the IAM role to use for this job request.

$sel:status:BatchImport', batchImport_status - The status of the batch import job.

$sel:processedRecordsCount:BatchImport', batchImport_processedRecordsCount - The number of records processed by batch import job.

$sel:totalRecordsCount:BatchImport', batchImport_totalRecordsCount - The total number of records in the batch import job.

$sel:jobId:BatchImport', batchImport_jobId - The ID of the batch import job.

$sel:arn:BatchImport', batchImport_arn - The ARN of the batch import job.

$sel:startTime:BatchImport', batchImport_startTime - Timestamp of when the batch import job started.

$sel:eventTypeName:BatchImport', batchImport_eventTypeName - The name of the event type.

$sel:completionTime:BatchImport', batchImport_completionTime - Timestamp of when batch import job completed.

$sel:outputPath:BatchImport', batchImport_outputPath - The Amazon S3 location of your output file.

$sel:inputPath:BatchImport', batchImport_inputPath - The Amazon S3 location of your data file for batch import.

$sel:failedRecordsCount:BatchImport', batchImport_failedRecordsCount - The number of records that failed to import.

batchImport_failureReason :: Lens' BatchImport (Maybe Text)

The reason batch import job failed.

batchImport_iamRoleArn :: Lens' BatchImport (Maybe Text)

The ARN of the IAM role to use for this job request.

batchImport_status :: Lens' BatchImport (Maybe AsyncJobStatus)

The status of the batch import job.

batchImport_processedRecordsCount :: Lens' BatchImport (Maybe Int)

The number of records processed by batch import job.

batchImport_totalRecordsCount :: Lens' BatchImport (Maybe Int)

The total number of records in the batch import job.

batchImport_jobId :: Lens' BatchImport (Maybe Text)

The ID of the batch import job.

batchImport_arn :: Lens' BatchImport (Maybe Text)

The ARN of the batch import job.

batchImport_startTime :: Lens' BatchImport (Maybe Text)

Timestamp of when the batch import job started.

batchImport_eventTypeName :: Lens' BatchImport (Maybe Text)

The name of the event type.

batchImport_completionTime :: Lens' BatchImport (Maybe Text)

Timestamp of when batch import job completed.

batchImport_outputPath :: Lens' BatchImport (Maybe Text)

The Amazon S3 location of your output file.

batchImport_inputPath :: Lens' BatchImport (Maybe Text)

The Amazon S3 location of your data file for batch import.

batchImport_failedRecordsCount :: Lens' BatchImport (Maybe Int)

The number of records that failed to import.