Network.AWS.FraudDetector.Types.ExternalModel

Description

Documentation

data ExternalModel

The Amazon SageMaker model.

See: newExternalModel smart constructor.

Constructors

ExternalModel'

Fields

modelEndpoint :: Maybe Text

The Amazon SageMaker model endpoints.

modelSource :: Maybe ModelSource

The source of the model.

lastUpdatedTime :: Maybe Text

Timestamp of when the model was last updated.

arn :: Maybe Text

The model ARN.

createdTime :: Maybe Text

Timestamp of when the model was last created.

modelEndpointStatus :: Maybe ModelEndpointStatus

The Amazon Fraud Detector status for the external model endpoint

outputConfiguration :: Maybe ModelOutputConfiguration

The output configuration.

invokeModelEndpointRoleArn :: Maybe Text

The role used to invoke the model.

inputConfiguration :: Maybe ModelInputConfiguration

The input configuration.

newExternalModel :: ExternalModel

Create a value of ExternalModel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:modelEndpoint:ExternalModel', externalModel_modelEndpoint - The Amazon SageMaker model endpoints.

$sel:modelSource:ExternalModel', externalModel_modelSource - The source of the model.

$sel:lastUpdatedTime:ExternalModel', externalModel_lastUpdatedTime - Timestamp of when the model was last updated.

$sel:arn:ExternalModel', externalModel_arn - The model ARN.

$sel:createdTime:ExternalModel', externalModel_createdTime - Timestamp of when the model was last created.

$sel:modelEndpointStatus:ExternalModel', externalModel_modelEndpointStatus - The Amazon Fraud Detector status for the external model endpoint

$sel:outputConfiguration:ExternalModel', externalModel_outputConfiguration - The output configuration.

$sel:invokeModelEndpointRoleArn:ExternalModel', externalModel_invokeModelEndpointRoleArn - The role used to invoke the model.

$sel:inputConfiguration:ExternalModel', externalModel_inputConfiguration - The input configuration.

externalModel_modelEndpoint :: Lens' ExternalModel (Maybe Text)

The Amazon SageMaker model endpoints.

externalModel_modelSource :: Lens' ExternalModel (Maybe ModelSource)

The source of the model.

externalModel_lastUpdatedTime :: Lens' ExternalModel (Maybe Text)

Timestamp of when the model was last updated.

externalModel_arn :: Lens' ExternalModel (Maybe Text)

The model ARN.

externalModel_createdTime :: Lens' ExternalModel (Maybe Text)

Timestamp of when the model was last created.

externalModel_modelEndpointStatus :: Lens' ExternalModel (Maybe ModelEndpointStatus)

The Amazon Fraud Detector status for the external model endpoint

externalModel_outputConfiguration :: Lens' ExternalModel (Maybe ModelOutputConfiguration)

The output configuration.

externalModel_invokeModelEndpointRoleArn :: Lens' ExternalModel (Maybe Text)

The role used to invoke the model.

externalModel_inputConfiguration :: Lens' ExternalModel (Maybe ModelInputConfiguration)

The input configuration.