Network.AWS.FraudDetector.Types.Model

Description

Documentation

data Model

The model.

See: newModel smart constructor.

Constructors

Model'

Fields

modelType :: Maybe ModelTypeEnum

The model type.

lastUpdatedTime :: Maybe Text

Timestamp of last time the model was updated.

modelId :: Maybe Text

The model ID.

arn :: Maybe Text

The ARN of the model.

createdTime :: Maybe Text

Timestamp of when the model was created.

eventTypeName :: Maybe Text

The name of the event type.

description :: Maybe Text

The model description.

newModel :: Model

Create a value of Model with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:modelType:Model', model_modelType - The model type.

$sel:lastUpdatedTime:Model', model_lastUpdatedTime - Timestamp of last time the model was updated.

$sel:modelId:Model', model_modelId - The model ID.

$sel:arn:Model', model_arn - The ARN of the model.

$sel:createdTime:Model', model_createdTime - Timestamp of when the model was created.

$sel:eventTypeName:Model', model_eventTypeName - The name of the event type.

$sel:description:Model', model_description - The model description.

model_modelType :: Lens' Model (Maybe ModelTypeEnum)

The model type.

model_lastUpdatedTime :: Lens' Model (Maybe Text)

Timestamp of last time the model was updated.

model_modelId :: Lens' Model (Maybe Text)

The model ID.

model_arn :: Lens' Model (Maybe Text)

The ARN of the model.

model_createdTime :: Lens' Model (Maybe Text)

Timestamp of when the model was created.

model_eventTypeName :: Lens' Model (Maybe Text)

The name of the event type.

model_description :: Lens' Model (Maybe Text)

The model description.