Network.AWS.FraudDetector.Types.Outcome

Description

Documentation

data Outcome

The outcome.

See: newOutcome smart constructor.

Constructors

Outcome'

Fields

lastUpdatedTime :: Maybe Text

The timestamp when the outcome was last updated.

arn :: Maybe Text

The outcome ARN.

createdTime :: Maybe Text

The timestamp when the outcome was created.

name :: Maybe Text

The outcome name.

description :: Maybe Text

The outcome description.

newOutcome :: Outcome

Create a value of Outcome with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastUpdatedTime:Outcome', outcome_lastUpdatedTime - The timestamp when the outcome was last updated.

$sel:arn:Outcome', outcome_arn - The outcome ARN.

$sel:createdTime:Outcome', outcome_createdTime - The timestamp when the outcome was created.

$sel:name:Outcome', outcome_name - The outcome name.

$sel:description:Outcome', outcome_description - The outcome description.

outcome_lastUpdatedTime :: Lens' Outcome (Maybe Text)

The timestamp when the outcome was last updated.

outcome_arn :: Lens' Outcome (Maybe Text)

The outcome ARN.

outcome_createdTime :: Lens' Outcome (Maybe Text)

The timestamp when the outcome was created.

outcome_name :: Lens' Outcome (Maybe Text)

The outcome name.

outcome_description :: Lens' Outcome (Maybe Text)

The outcome description.