Network.AWS.FraudDetector.Types.Variable

Description

Documentation

data Variable

The variable.

See: newVariable smart constructor.

Constructors

Variable'

Fields

lastUpdatedTime :: Maybe Text

The time when variable was last updated.

arn :: Maybe Text

The ARN of the variable.

createdTime :: Maybe Text

The time when the variable was created.

name :: Maybe Text

The name of the variable.

dataSource :: Maybe DataSource

The data source of the variable.

dataType :: Maybe DataType

The data type of the variable. For more information see Variable types.

defaultValue :: Maybe Text

The default value of the variable.

variableType :: Maybe Text

The variable type of the variable.

Valid Values: AUTH_CODE | AVS | BILLING_ADDRESS_L1 | BILLING_ADDRESS_L2 | BILLING_CITY | BILLING_COUNTRY | BILLING_NAME | BILLING_PHONE | BILLING_STATE | BILLING_ZIP | CARD_BIN | CATEGORICAL | CURRENCY_CODE | EMAIL_ADDRESS | FINGERPRINT | FRAUD_LABEL | FREE_FORM_TEXT | IP_ADDRESS | NUMERIC | ORDER_ID | PAYMENT_TYPE | PHONE_NUMBER | PRICE | PRODUCT_CATEGORY | SHIPPING_ADDRESS_L1 | SHIPPING_ADDRESS_L2 | SHIPPING_CITY | SHIPPING_COUNTRY | SHIPPING_NAME | SHIPPING_PHONE | SHIPPING_STATE | SHIPPING_ZIP | USERAGENT

description :: Maybe Text

The description of the variable.

newVariable :: Variable

Create a value of Variable with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastUpdatedTime:Variable', variable_lastUpdatedTime - The time when variable was last updated.

$sel:arn:Variable', variable_arn - The ARN of the variable.

$sel:createdTime:Variable', variable_createdTime - The time when the variable was created.

$sel:name:Variable', variable_name - The name of the variable.

$sel:dataSource:Variable', variable_dataSource - The data source of the variable.

$sel:dataType:Variable', variable_dataType - The data type of the variable. For more information see Variable types.

$sel:defaultValue:Variable', variable_defaultValue - The default value of the variable.

$sel:variableType:Variable', variable_variableType - The variable type of the variable.

Valid Values: AUTH_CODE | AVS | BILLING_ADDRESS_L1 | BILLING_ADDRESS_L2 | BILLING_CITY | BILLING_COUNTRY | BILLING_NAME | BILLING_PHONE | BILLING_STATE | BILLING_ZIP | CARD_BIN | CATEGORICAL | CURRENCY_CODE | EMAIL_ADDRESS | FINGERPRINT | FRAUD_LABEL | FREE_FORM_TEXT | IP_ADDRESS | NUMERIC | ORDER_ID | PAYMENT_TYPE | PHONE_NUMBER | PRICE | PRODUCT_CATEGORY | SHIPPING_ADDRESS_L1 | SHIPPING_ADDRESS_L2 | SHIPPING_CITY | SHIPPING_COUNTRY | SHIPPING_NAME | SHIPPING_PHONE | SHIPPING_STATE | SHIPPING_ZIP | USERAGENT

$sel:description:Variable', variable_description - The description of the variable.

variable_lastUpdatedTime :: Lens' Variable (Maybe Text)

The time when variable was last updated.

variable_arn :: Lens' Variable (Maybe Text)

The ARN of the variable.

variable_createdTime :: Lens' Variable (Maybe Text)

The time when the variable was created.

variable_name :: Lens' Variable (Maybe Text)

The name of the variable.

variable_dataSource :: Lens' Variable (Maybe DataSource)

The data source of the variable.

variable_dataType :: Lens' Variable (Maybe DataType)

The data type of the variable. For more information see Variable types.

variable_defaultValue :: Lens' Variable (Maybe Text)

The default value of the variable.

variable_variableType :: Lens' Variable (Maybe Text)

The variable type of the variable.

Valid Values: AUTH_CODE | AVS | BILLING_ADDRESS_L1 | BILLING_ADDRESS_L2 | BILLING_CITY | BILLING_COUNTRY | BILLING_NAME | BILLING_PHONE | BILLING_STATE | BILLING_ZIP | CARD_BIN | CATEGORICAL | CURRENCY_CODE | EMAIL_ADDRESS | FINGERPRINT | FRAUD_LABEL | FREE_FORM_TEXT | IP_ADDRESS | NUMERIC | ORDER_ID | PAYMENT_TYPE | PHONE_NUMBER | PRICE | PRODUCT_CATEGORY | SHIPPING_ADDRESS_L1 | SHIPPING_ADDRESS_L2 | SHIPPING_CITY | SHIPPING_COUNTRY | SHIPPING_NAME | SHIPPING_PHONE | SHIPPING_STATE | SHIPPING_ZIP | USERAGENT

variable_description :: Lens' Variable (Maybe Text)

The description of the variable.