Network.AWS.FraudDetector.UpdateRuleVersion

Description

Updates a rule version resulting in a new rule version. Updates a rule version resulting in a new rule version (version 1, 2, 3 ...).

Creating a Request

data UpdateRuleVersion

See: newUpdateRuleVersion smart constructor.

Constructors

UpdateRuleVersion'

Fields

description :: Maybe Text

The description.

tags :: Maybe [Tag]

The tags to assign to the rule version.

rule :: Rule

The rule to update.

expression :: Sensitive Text

The rule expression.

language :: Language

The language.

outcomes :: NonEmpty Text

The outcomes.

newUpdateRuleVersion

Arguments

:: Rule
$sel:rule:UpdateRuleVersion'

-> Text
$sel:expression:UpdateRuleVersion'

-> Language
$sel:language:UpdateRuleVersion'

-> NonEmpty Text
$sel:outcomes:UpdateRuleVersion'

-> UpdateRuleVersion

Create a value of UpdateRuleVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:UpdateRuleVersion', updateRuleVersion_description - The description.

$sel:tags:UpdateRuleVersion', updateRuleVersion_tags - The tags to assign to the rule version.

$sel:rule:UpdateRuleVersion', updateRuleVersion_rule - The rule to update.

$sel:expression:UpdateRuleVersion', updateRuleVersion_expression - The rule expression.

$sel:language:UpdateRuleVersion', updateRuleVersion_language - The language.

$sel:outcomes:UpdateRuleVersion', updateRuleVersion_outcomes - The outcomes.

Request Lenses

updateRuleVersion_description :: Lens' UpdateRuleVersion (Maybe Text)

The description.

updateRuleVersion_tags :: Lens' UpdateRuleVersion (Maybe [Tag])

The tags to assign to the rule version.

updateRuleVersion_rule :: Lens' UpdateRuleVersion Rule

The rule to update.

updateRuleVersion_expression :: Lens' UpdateRuleVersion Text

The rule expression.

updateRuleVersion_language :: Lens' UpdateRuleVersion Language

The language.

updateRuleVersion_outcomes :: Lens' UpdateRuleVersion (NonEmpty Text)

The outcomes.

Destructuring the Response

data UpdateRuleVersionResponse

See: newUpdateRuleVersionResponse smart constructor.

Constructors

UpdateRuleVersionResponse'

Fields

rule :: Maybe Rule

The new rule version that was created.

httpStatus :: Int

The response's http status code.

newUpdateRuleVersionResponse

Arguments

:: Int
$sel:httpStatus:UpdateRuleVersionResponse'

-> UpdateRuleVersionResponse

Create a value of UpdateRuleVersionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:rule:UpdateRuleVersion', updateRuleVersionResponse_rule - The new rule version that was created.

$sel:httpStatus:UpdateRuleVersionResponse', updateRuleVersionResponse_httpStatus - The response's http status code.

Response Lenses

updateRuleVersionResponse_rule :: Lens' UpdateRuleVersionResponse (Maybe Rule)

The new rule version that was created.

updateRuleVersionResponse_httpStatus :: Lens' UpdateRuleVersionResponse Int

The response's http status code.