Network.AWS.GameLift.CreateBuild

Description

Creates a new Amazon GameLift build resource for your game server binary files. Game server binaries must be combined into a zip file for use with Amazon GameLift.

When setting up a new game build for GameLift, we recommend using the AWS CLI command upload-build . This helper command combines two tasks: (1) it uploads your build files from a file directory to a GameLift Amazon S3 location, and (2) it creates a new build resource.

The CreateBuild operation can used in the following scenarios:

To create a new game build with build files that are in an Amazon S3 location under an AWS account that you control. To use this option, you must first give Amazon GameLift access to the Amazon S3 bucket. With permissions in place, call CreateBuild and specify a build name, operating system, and the Amazon S3 storage location of your game build.

To directly upload your build files to a GameLift Amazon S3 location. To use this option, first call CreateBuild and specify a build name and operating system. This operation creates a new build resource and also returns an Amazon S3 location with temporary access credentials. Use the credentials to manually upload your build files to the specified Amazon S3 location. For more information, see Uploading Objects in the Amazon S3 Developer Guide. Build files can be uploaded to the GameLift Amazon S3 location once only; that can't be updated.

If successful, this operation creates a new build resource with a unique build ID and places it in INITIALIZED status. A build must be in READY status before you can create fleets with it.

Learn more

Uploading Your Game

Create a Build with Files in Amazon S3

Related actions

CreateBuild | ListBuilds | DescribeBuild | UpdateBuild | DeleteBuild | All APIs by task

Creating a Request

data CreateBuild

Represents the input for a request operation.

See: newCreateBuild smart constructor.

Constructors

CreateBuild'

Fields

version :: Maybe Text

Version information that is associated with a build or script. Version strings do not need to be unique. You can use UpdateBuild to change this value later.

name :: Maybe Text

A descriptive label that is associated with a build. Build names do not need to be unique. You can use UpdateBuild to change this value later.

storageLocation :: Maybe S3Location

Information indicating where your game build files are stored. Use this parameter only when creating a build with files stored in an Amazon S3 bucket that you own. The storage location must specify an Amazon S3 bucket name and key. The location must also specify a role ARN that you set up to allow Amazon GameLift to access your Amazon S3 bucket. The S3 bucket and your new build must be in the same Region.

tags :: Maybe [Tag]

A list of labels to assign to the new build resource. Tags are developer-defined key-value pairs. Tagging AWS resources are useful for resource management, access management and cost allocation. For more information, see Tagging AWS Resources in the AWS General Reference. Once the resource is created, you can use TagResource, UntagResource, and ListTagsForResource to add, remove, and view tags. The maximum tag limit may be lower than stated. See the AWS General Reference for actual tagging limits.

operatingSystem :: Maybe OperatingSystem

The operating system that the game server binaries are built to run on. This value determines the type of fleet resources that you can use for this build. If your game build contains multiple executables, they all must run on the same operating system. If an operating system is not specified when creating a build, Amazon GameLift uses the default value (WINDOWS_2012). This value cannot be changed later.

newCreateBuild :: CreateBuild

Create a value of CreateBuild with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:version:CreateBuild', createBuild_version - Version information that is associated with a build or script. Version strings do not need to be unique. You can use UpdateBuild to change this value later.

$sel:name:CreateBuild', createBuild_name - A descriptive label that is associated with a build. Build names do not need to be unique. You can use UpdateBuild to change this value later.

$sel:storageLocation:CreateBuild', createBuild_storageLocation - Information indicating where your game build files are stored. Use this parameter only when creating a build with files stored in an Amazon S3 bucket that you own. The storage location must specify an Amazon S3 bucket name and key. The location must also specify a role ARN that you set up to allow Amazon GameLift to access your Amazon S3 bucket. The S3 bucket and your new build must be in the same Region.

$sel:tags:CreateBuild', createBuild_tags - A list of labels to assign to the new build resource. Tags are developer-defined key-value pairs. Tagging AWS resources are useful for resource management, access management and cost allocation. For more information, see Tagging AWS Resources in the AWS General Reference. Once the resource is created, you can use TagResource, UntagResource, and ListTagsForResource to add, remove, and view tags. The maximum tag limit may be lower than stated. See the AWS General Reference for actual tagging limits.

$sel:operatingSystem:CreateBuild', createBuild_operatingSystem - The operating system that the game server binaries are built to run on. This value determines the type of fleet resources that you can use for this build. If your game build contains multiple executables, they all must run on the same operating system. If an operating system is not specified when creating a build, Amazon GameLift uses the default value (WINDOWS_2012). This value cannot be changed later.

Request Lenses

createBuild_version :: Lens' CreateBuild (Maybe Text)

Version information that is associated with a build or script. Version strings do not need to be unique. You can use UpdateBuild to change this value later.

createBuild_name :: Lens' CreateBuild (Maybe Text)

A descriptive label that is associated with a build. Build names do not need to be unique. You can use UpdateBuild to change this value later.

createBuild_storageLocation :: Lens' CreateBuild (Maybe S3Location)

Information indicating where your game build files are stored. Use this parameter only when creating a build with files stored in an Amazon S3 bucket that you own. The storage location must specify an Amazon S3 bucket name and key. The location must also specify a role ARN that you set up to allow Amazon GameLift to access your Amazon S3 bucket. The S3 bucket and your new build must be in the same Region.

createBuild_tags :: Lens' CreateBuild (Maybe [Tag])

A list of labels to assign to the new build resource. Tags are developer-defined key-value pairs. Tagging AWS resources are useful for resource management, access management and cost allocation. For more information, see Tagging AWS Resources in the AWS General Reference. Once the resource is created, you can use TagResource, UntagResource, and ListTagsForResource to add, remove, and view tags. The maximum tag limit may be lower than stated. See the AWS General Reference for actual tagging limits.

createBuild_operatingSystem :: Lens' CreateBuild (Maybe OperatingSystem)

The operating system that the game server binaries are built to run on. This value determines the type of fleet resources that you can use for this build. If your game build contains multiple executables, they all must run on the same operating system. If an operating system is not specified when creating a build, Amazon GameLift uses the default value (WINDOWS_2012). This value cannot be changed later.

Destructuring the Response

data CreateBuildResponse

Represents the returned data in response to a request operation.

See: newCreateBuildResponse smart constructor.

Constructors

CreateBuildResponse'

Fields

build :: Maybe Build

The newly created build resource, including a unique build IDs and status.

storageLocation :: Maybe S3Location

Amazon S3 location for your game build file, including bucket name and key.

uploadCredentials :: Maybe (Sensitive AwsCredentials)

This element is returned only when the operation is called without a storage location. It contains credentials to use when you are uploading a build file to an Amazon S3 bucket that is owned by Amazon GameLift. Credentials have a limited life span. To refresh these credentials, call RequestUploadCredentials.

httpStatus :: Int

The response's http status code.

newCreateBuildResponse

Arguments

:: Int
$sel:httpStatus:CreateBuildResponse'

-> CreateBuildResponse

Create a value of CreateBuildResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:build:CreateBuildResponse', createBuildResponse_build - The newly created build resource, including a unique build IDs and status.

$sel:storageLocation:CreateBuild', createBuildResponse_storageLocation - Amazon S3 location for your game build file, including bucket name and key.

$sel:uploadCredentials:CreateBuildResponse', createBuildResponse_uploadCredentials - This element is returned only when the operation is called without a storage location. It contains credentials to use when you are uploading a build file to an Amazon S3 bucket that is owned by Amazon GameLift. Credentials have a limited life span. To refresh these credentials, call RequestUploadCredentials.

$sel:httpStatus:CreateBuildResponse', createBuildResponse_httpStatus - The response's http status code.

Response Lenses

createBuildResponse_build :: Lens' CreateBuildResponse (Maybe Build)

The newly created build resource, including a unique build IDs and status.

createBuildResponse_storageLocation :: Lens' CreateBuildResponse (Maybe S3Location)

Amazon S3 location for your game build file, including bucket name and key.

createBuildResponse_uploadCredentials :: Lens' CreateBuildResponse (Maybe AwsCredentials)

This element is returned only when the operation is called without a storage location. It contains credentials to use when you are uploading a build file to an Amazon S3 bucket that is owned by Amazon GameLift. Credentials have a limited life span. To refresh these credentials, call RequestUploadCredentials.

createBuildResponse_httpStatus :: Lens' CreateBuildResponse Int

The response's http status code.