Network.AWS.GameLift.CreateGameSessionQueue

Description

Creates a placement queue that processes requests for new game sessions. A queue uses FleetIQ algorithms to determine the best placement locations and find an available game server there, then prompts the game server process to start a new game session.

A game session queue is configured with a set of destinations (GameLift fleets or aliases), which determine the locations where the queue can place new game sessions. These destinations can span multiple fleet types (Spot and On-Demand), instance types, and AWS Regions. If the queue includes multi-location fleets, the queue is able to place game sessions in all of a fleet's remote locations. You can opt to filter out individual locations if needed.

The queue configuration also determines how FleetIQ selects the best available placement for a new game session. Before searching for an available game server, FleetIQ first prioritizes the queue's destinations and locations, with the best placement locations on top. You can set up the queue to use the FleetIQ default prioritization or provide an alternate set of priorities.

To create a new queue, provide a name, timeout value, and a list of destinations. Optionally, specify a sort configuration and/or a filter, and define a set of latency cap policies. You can also include the ARN for an Amazon Simple Notification Service (SNS) topic to receive notifications of game session placement activity. Notifications using SNS or CloudWatch events is the preferred way to track placement activity.

If successful, a new GameSessionQueue object is returned with an assigned queue ARN. New game session requests, which are submitted to the queue with StartGameSessionPlacement or StartMatchmaking, reference a queue's name or ARN.

Learn more

Design a game session queue

Create a game session queue

Related actions

CreateGameSessionQueue | DescribeGameSessionQueues | UpdateGameSessionQueue | DeleteGameSessionQueue | All APIs by task

Creating a Request

data CreateGameSessionQueue

Represents the input for a request operation.

See: newCreateGameSessionQueue smart constructor.

Constructors

CreateGameSessionQueue'

Fields

playerLatencyPolicies :: Maybe [PlayerLatencyPolicy]

A set of policies that act as a sliding cap on player latency. FleetIQ works to deliver low latency for most players in a game session. These policies ensure that no individual player can be placed into a game with unreasonably high latency. Use multiple policies to gradually relax latency requirements a step at a time. Multiple policies are applied based on their maximum allowed latency, starting with the lowest value.

filterConfiguration :: Maybe FilterConfiguration

A list of locations where a queue is allowed to place new game sessions. Locations are specified in the form of AWS Region codes, such as us-west-2. If this parameter is not set, game sessions can be placed in any queue location.

notificationTarget :: Maybe Text

An SNS topic ARN that is set up to receive game session placement notifications. See Setting up notifications for game session placement.

timeoutInSeconds :: Maybe Natural

The maximum time, in seconds, that a new game session placement request remains in the queue. When a request exceeds this time, the game session placement changes to a TIMED_OUT status.

destinations :: Maybe [GameSessionQueueDestination]

A list of fleets and/or fleet aliases that can be used to fulfill game session placement requests in the queue. Destinations are identified by either a fleet ARN or a fleet alias ARN, and are listed in order of placement preference.

customEventData :: Maybe Text

Information to be added to all events that are related to this game session queue.

priorityConfiguration :: Maybe PriorityConfiguration

Custom settings to use when prioritizing destinations and locations for game session placements. This configuration replaces the FleetIQ default prioritization process. Priority types that are not explicitly named will be automatically applied at the end of the prioritization process.

tags :: Maybe [Tag]

A list of labels to assign to the new game session queue resource. Tags are developer-defined key-value pairs. Tagging AWS resources are useful for resource management, access management and cost allocation. For more information, see Tagging AWS Resources in the AWS General Reference. Once the resource is created, you can use TagResource, UntagResource, and ListTagsForResource to add, remove, and view tags. The maximum tag limit may be lower than stated. See the AWS General Reference for actual tagging limits.

name :: Text

A descriptive label that is associated with game session queue. Queue names must be unique within each Region.

newCreateGameSessionQueue

Arguments

:: Text
$sel:name:CreateGameSessionQueue'

-> CreateGameSessionQueue

Create a value of CreateGameSessionQueue with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:playerLatencyPolicies:CreateGameSessionQueue', createGameSessionQueue_playerLatencyPolicies - A set of policies that act as a sliding cap on player latency. FleetIQ works to deliver low latency for most players in a game session. These policies ensure that no individual player can be placed into a game with unreasonably high latency. Use multiple policies to gradually relax latency requirements a step at a time. Multiple policies are applied based on their maximum allowed latency, starting with the lowest value.

$sel:filterConfiguration:CreateGameSessionQueue', createGameSessionQueue_filterConfiguration - A list of locations where a queue is allowed to place new game sessions. Locations are specified in the form of AWS Region codes, such as us-west-2. If this parameter is not set, game sessions can be placed in any queue location.

$sel:notificationTarget:CreateGameSessionQueue', createGameSessionQueue_notificationTarget - An SNS topic ARN that is set up to receive game session placement notifications. See Setting up notifications for game session placement.

$sel:timeoutInSeconds:CreateGameSessionQueue', createGameSessionQueue_timeoutInSeconds - The maximum time, in seconds, that a new game session placement request remains in the queue. When a request exceeds this time, the game session placement changes to a TIMED_OUT status.

$sel:destinations:CreateGameSessionQueue', createGameSessionQueue_destinations - A list of fleets and/or fleet aliases that can be used to fulfill game session placement requests in the queue. Destinations are identified by either a fleet ARN or a fleet alias ARN, and are listed in order of placement preference.

$sel:customEventData:CreateGameSessionQueue', createGameSessionQueue_customEventData - Information to be added to all events that are related to this game session queue.

$sel:priorityConfiguration:CreateGameSessionQueue', createGameSessionQueue_priorityConfiguration - Custom settings to use when prioritizing destinations and locations for game session placements. This configuration replaces the FleetIQ default prioritization process. Priority types that are not explicitly named will be automatically applied at the end of the prioritization process.

$sel:tags:CreateGameSessionQueue', createGameSessionQueue_tags - A list of labels to assign to the new game session queue resource. Tags are developer-defined key-value pairs. Tagging AWS resources are useful for resource management, access management and cost allocation. For more information, see Tagging AWS Resources in the AWS General Reference. Once the resource is created, you can use TagResource, UntagResource, and ListTagsForResource to add, remove, and view tags. The maximum tag limit may be lower than stated. See the AWS General Reference for actual tagging limits.

$sel:name:CreateGameSessionQueue', createGameSessionQueue_name - A descriptive label that is associated with game session queue. Queue names must be unique within each Region.

Request Lenses

createGameSessionQueue_playerLatencyPolicies :: Lens' CreateGameSessionQueue (Maybe [PlayerLatencyPolicy])

A set of policies that act as a sliding cap on player latency. FleetIQ works to deliver low latency for most players in a game session. These policies ensure that no individual player can be placed into a game with unreasonably high latency. Use multiple policies to gradually relax latency requirements a step at a time. Multiple policies are applied based on their maximum allowed latency, starting with the lowest value.

createGameSessionQueue_filterConfiguration :: Lens' CreateGameSessionQueue (Maybe FilterConfiguration)

A list of locations where a queue is allowed to place new game sessions. Locations are specified in the form of AWS Region codes, such as us-west-2. If this parameter is not set, game sessions can be placed in any queue location.

createGameSessionQueue_notificationTarget :: Lens' CreateGameSessionQueue (Maybe Text)

An SNS topic ARN that is set up to receive game session placement notifications. See Setting up notifications for game session placement.

createGameSessionQueue_timeoutInSeconds :: Lens' CreateGameSessionQueue (Maybe Natural)

The maximum time, in seconds, that a new game session placement request remains in the queue. When a request exceeds this time, the game session placement changes to a TIMED_OUT status.

createGameSessionQueue_destinations :: Lens' CreateGameSessionQueue (Maybe [GameSessionQueueDestination])

A list of fleets and/or fleet aliases that can be used to fulfill game session placement requests in the queue. Destinations are identified by either a fleet ARN or a fleet alias ARN, and are listed in order of placement preference.

createGameSessionQueue_customEventData :: Lens' CreateGameSessionQueue (Maybe Text)

Information to be added to all events that are related to this game session queue.

createGameSessionQueue_priorityConfiguration :: Lens' CreateGameSessionQueue (Maybe PriorityConfiguration)

Custom settings to use when prioritizing destinations and locations for game session placements. This configuration replaces the FleetIQ default prioritization process. Priority types that are not explicitly named will be automatically applied at the end of the prioritization process.

createGameSessionQueue_tags :: Lens' CreateGameSessionQueue (Maybe [Tag])

A list of labels to assign to the new game session queue resource. Tags are developer-defined key-value pairs. Tagging AWS resources are useful for resource management, access management and cost allocation. For more information, see Tagging AWS Resources in the AWS General Reference. Once the resource is created, you can use TagResource, UntagResource, and ListTagsForResource to add, remove, and view tags. The maximum tag limit may be lower than stated. See the AWS General Reference for actual tagging limits.

createGameSessionQueue_name :: Lens' CreateGameSessionQueue Text

A descriptive label that is associated with game session queue. Queue names must be unique within each Region.

Destructuring the Response

data CreateGameSessionQueueResponse

Represents the returned data in response to a request operation.

See: newCreateGameSessionQueueResponse smart constructor.

Constructors

CreateGameSessionQueueResponse'

Fields

gameSessionQueue :: Maybe GameSessionQueue

An object that describes the newly created game session queue.

httpStatus :: Int

The response's http status code.

newCreateGameSessionQueueResponse

Arguments

:: Int
$sel:httpStatus:CreateGameSessionQueueResponse'

-> CreateGameSessionQueueResponse

Create a value of CreateGameSessionQueueResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:gameSessionQueue:CreateGameSessionQueueResponse', createGameSessionQueueResponse_gameSessionQueue - An object that describes the newly created game session queue.

$sel:httpStatus:CreateGameSessionQueueResponse', createGameSessionQueueResponse_httpStatus - The response's http status code.

Response Lenses

createGameSessionQueueResponse_gameSessionQueue :: Lens' CreateGameSessionQueueResponse (Maybe GameSessionQueue)

An object that describes the newly created game session queue.

createGameSessionQueueResponse_httpStatus :: Lens' CreateGameSessionQueueResponse Int

The response's http status code.