Network.AWS.GameLift.CreateMatchmakingConfiguration

Description

Defines a new matchmaking configuration for use with FlexMatch. Whether your are using FlexMatch with GameLift hosting or as a standalone matchmaking service, the matchmaking configuration sets out rules for matching players and forming teams. If you're also using GameLift hosting, it defines how to start game sessions for each match. Your matchmaking system can use multiple configurations to handle different game scenarios. All matchmaking requests (StartMatchmaking or StartMatchBackfill) identify the matchmaking configuration to use and provide player attributes consistent with that configuration.

To create a matchmaking configuration, you must provide the following: configuration name and FlexMatch mode (with or without GameLift hosting); a rule set that specifies how to evaluate players and find acceptable matches; whether player acceptance is required; and the maximum time allowed for a matchmaking attempt. When using FlexMatch with GameLift hosting, you also need to identify the game session queue to use when starting a game session for the match.

In addition, you must set up an Amazon Simple Notification Service (SNS) topic to receive matchmaking notifications. Provide the topic ARN in the matchmaking configuration. An alternative method, continuously polling ticket status with DescribeMatchmaking, is only suitable for games in development with low matchmaking usage.

Learn more

Design a FlexMatch matchmaker

Set up FlexMatch event notification

Related actions

CreateMatchmakingConfiguration | DescribeMatchmakingConfigurations | UpdateMatchmakingConfiguration | DeleteMatchmakingConfiguration | CreateMatchmakingRuleSet | DescribeMatchmakingRuleSets | ValidateMatchmakingRuleSet | DeleteMatchmakingRuleSet | All APIs by task

Creating a Request

data CreateMatchmakingConfiguration

Represents the input for a request operation.

See: newCreateMatchmakingConfiguration smart constructor.

Constructors

CreateMatchmakingConfiguration'

Fields

flexMatchMode :: Maybe FlexMatchMode

Indicates whether this matchmaking configuration is being used with GameLift hosting or as a standalone matchmaking solution.

STANDALONE - FlexMatch forms matches and returns match information, including players and team assignments, in a MatchmakingSucceeded event.

WITH_QUEUE - FlexMatch forms matches and uses the specified GameLift queue to start a game session for the match.

customEventData :: Maybe Text

Information to be added to all events related to this matchmaking configuration.

backfillMode :: Maybe BackfillMode

The method used to backfill game sessions that are created with this matchmaking configuration. Specify MANUAL when your game manages backfill requests manually or does not use the match backfill feature. Specify AUTOMATIC to have GameLift create a StartMatchBackfill request whenever a game session has one or more open slots. Learn more about manual and automatic backfill in Backfill Existing Games with FlexMatch. Automatic backfill is not available when FlexMatchMode is set to STANDALONE.

gameProperties :: Maybe [GameProperty]

A set of custom properties for a game session, formatted as key:value pairs. These properties are passed to a game server process in the GameSession object with a request to start a new game session (see Start a Game Session). This information is added to the new GameSession object that is created for a successful match. This parameter is not used if FlexMatchMode is set to STANDALONE.

acceptanceTimeoutSeconds :: Maybe Natural

The length of time (in seconds) to wait for players to accept a proposed match, if acceptance is required.

additionalPlayerCount :: Maybe Natural

The number of player slots in a match to keep open for future players. For example, if the configuration's rule set specifies a match for a single 12-person team, and the additional player count is set to 2, only 10 players are selected for the match. This parameter is not used if FlexMatchMode is set to STANDALONE.

gameSessionData :: Maybe Text

A set of custom game session properties, formatted as a single string value. This data is passed to a game server process in the GameSession object with a request to start a new game session (see Start a Game Session). This information is added to the new GameSession object that is created for a successful match. This parameter is not used if FlexMatchMode is set to STANDALONE.

gameSessionQueueArns :: Maybe [Text]

The Amazon Resource Name (ARN) that is assigned to a GameLift game session queue resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:<region>::gamesessionqueue/<queue name>. Queues can be located in any Region. Queues are used to start new GameLift-hosted game sessions for matches that are created with this matchmaking configuration. If FlexMatchMode is set to STANDALONE, do not set this parameter.

notificationTarget :: Maybe Text

An SNS topic ARN that is set up to receive matchmaking notifications. See Setting up notifications for matchmaking for more information.

tags :: Maybe [Tag]

A list of labels to assign to the new matchmaking configuration resource. Tags are developer-defined key-value pairs. Tagging AWS resources are useful for resource management, access management and cost allocation. For more information, see Tagging AWS Resources in the AWS General Reference. Once the resource is created, you can use TagResource, UntagResource, and ListTagsForResource to add, remove, and view tags. The maximum tag limit may be lower than stated. See the AWS General Reference for actual tagging limits.

description :: Maybe Text

A human-readable description of the matchmaking configuration.

name :: Text

A unique identifier for the matchmaking configuration. This name is used to identify the configuration associated with a matchmaking request or ticket.

requestTimeoutSeconds :: Natural

The maximum duration, in seconds, that a matchmaking ticket can remain in process before timing out. Requests that fail due to timing out can be resubmitted as needed.

acceptanceRequired :: Bool

A flag that determines whether a match that was created with this configuration must be accepted by the matched players. To require acceptance, set to TRUE. With this option enabled, matchmaking tickets use the status REQUIRES_ACCEPTANCE to indicate when a completed potential match is waiting for player acceptance.

ruleSetName :: Text

A unique identifier for the matchmaking rule set to use with this configuration. You can use either the rule set name or ARN value. A matchmaking configuration can only use rule sets that are defined in the same Region.

newCreateMatchmakingConfiguration

Arguments

:: Text
$sel:name:CreateMatchmakingConfiguration'

-> Natural
$sel:requestTimeoutSeconds:CreateMatchmakingConfiguration'

-> Bool
$sel:acceptanceRequired:CreateMatchmakingConfiguration'

-> Text
$sel:ruleSetName:CreateMatchmakingConfiguration'

-> CreateMatchmakingConfiguration

Create a value of CreateMatchmakingConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:flexMatchMode:CreateMatchmakingConfiguration', createMatchmakingConfiguration_flexMatchMode - Indicates whether this matchmaking configuration is being used with GameLift hosting or as a standalone matchmaking solution.

STANDALONE - FlexMatch forms matches and returns match information, including players and team assignments, in a MatchmakingSucceeded event.

WITH_QUEUE - FlexMatch forms matches and uses the specified GameLift queue to start a game session for the match.

$sel:customEventData:CreateMatchmakingConfiguration', createMatchmakingConfiguration_customEventData - Information to be added to all events related to this matchmaking configuration.

$sel:backfillMode:CreateMatchmakingConfiguration', createMatchmakingConfiguration_backfillMode - The method used to backfill game sessions that are created with this matchmaking configuration. Specify MANUAL when your game manages backfill requests manually or does not use the match backfill feature. Specify AUTOMATIC to have GameLift create a StartMatchBackfill request whenever a game session has one or more open slots. Learn more about manual and automatic backfill in Backfill Existing Games with FlexMatch. Automatic backfill is not available when FlexMatchMode is set to STANDALONE.

$sel:gameProperties:CreateMatchmakingConfiguration', createMatchmakingConfiguration_gameProperties - A set of custom properties for a game session, formatted as key:value pairs. These properties are passed to a game server process in the GameSession object with a request to start a new game session (see Start a Game Session). This information is added to the new GameSession object that is created for a successful match. This parameter is not used if FlexMatchMode is set to STANDALONE.

$sel:acceptanceTimeoutSeconds:CreateMatchmakingConfiguration', createMatchmakingConfiguration_acceptanceTimeoutSeconds - The length of time (in seconds) to wait for players to accept a proposed match, if acceptance is required.

$sel:additionalPlayerCount:CreateMatchmakingConfiguration', createMatchmakingConfiguration_additionalPlayerCount - The number of player slots in a match to keep open for future players. For example, if the configuration's rule set specifies a match for a single 12-person team, and the additional player count is set to 2, only 10 players are selected for the match. This parameter is not used if FlexMatchMode is set to STANDALONE.

$sel:gameSessionData:CreateMatchmakingConfiguration', createMatchmakingConfiguration_gameSessionData - A set of custom game session properties, formatted as a single string value. This data is passed to a game server process in the GameSession object with a request to start a new game session (see Start a Game Session). This information is added to the new GameSession object that is created for a successful match. This parameter is not used if FlexMatchMode is set to STANDALONE.

$sel:gameSessionQueueArns:CreateMatchmakingConfiguration', createMatchmakingConfiguration_gameSessionQueueArns - The Amazon Resource Name (ARN) that is assigned to a GameLift game session queue resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:<region>::gamesessionqueue/<queue name>. Queues can be located in any Region. Queues are used to start new GameLift-hosted game sessions for matches that are created with this matchmaking configuration. If FlexMatchMode is set to STANDALONE, do not set this parameter.

$sel:notificationTarget:CreateMatchmakingConfiguration', createMatchmakingConfiguration_notificationTarget - An SNS topic ARN that is set up to receive matchmaking notifications. See Setting up notifications for matchmaking for more information.

$sel:tags:CreateMatchmakingConfiguration', createMatchmakingConfiguration_tags - A list of labels to assign to the new matchmaking configuration resource. Tags are developer-defined key-value pairs. Tagging AWS resources are useful for resource management, access management and cost allocation. For more information, see Tagging AWS Resources in the AWS General Reference. Once the resource is created, you can use TagResource, UntagResource, and ListTagsForResource to add, remove, and view tags. The maximum tag limit may be lower than stated. See the AWS General Reference for actual tagging limits.

$sel:description:CreateMatchmakingConfiguration', createMatchmakingConfiguration_description - A human-readable description of the matchmaking configuration.

$sel:name:CreateMatchmakingConfiguration', createMatchmakingConfiguration_name - A unique identifier for the matchmaking configuration. This name is used to identify the configuration associated with a matchmaking request or ticket.

$sel:requestTimeoutSeconds:CreateMatchmakingConfiguration', createMatchmakingConfiguration_requestTimeoutSeconds - The maximum duration, in seconds, that a matchmaking ticket can remain in process before timing out. Requests that fail due to timing out can be resubmitted as needed.

$sel:acceptanceRequired:CreateMatchmakingConfiguration', createMatchmakingConfiguration_acceptanceRequired - A flag that determines whether a match that was created with this configuration must be accepted by the matched players. To require acceptance, set to TRUE. With this option enabled, matchmaking tickets use the status REQUIRES_ACCEPTANCE to indicate when a completed potential match is waiting for player acceptance.

$sel:ruleSetName:CreateMatchmakingConfiguration', createMatchmakingConfiguration_ruleSetName - A unique identifier for the matchmaking rule set to use with this configuration. You can use either the rule set name or ARN value. A matchmaking configuration can only use rule sets that are defined in the same Region.

Request Lenses

createMatchmakingConfiguration_flexMatchMode :: Lens' CreateMatchmakingConfiguration (Maybe FlexMatchMode)

Indicates whether this matchmaking configuration is being used with GameLift hosting or as a standalone matchmaking solution.

STANDALONE - FlexMatch forms matches and returns match information, including players and team assignments, in a MatchmakingSucceeded event.

WITH_QUEUE - FlexMatch forms matches and uses the specified GameLift queue to start a game session for the match.

createMatchmakingConfiguration_customEventData :: Lens' CreateMatchmakingConfiguration (Maybe Text)

Information to be added to all events related to this matchmaking configuration.

createMatchmakingConfiguration_backfillMode :: Lens' CreateMatchmakingConfiguration (Maybe BackfillMode)

The method used to backfill game sessions that are created with this matchmaking configuration. Specify MANUAL when your game manages backfill requests manually or does not use the match backfill feature. Specify AUTOMATIC to have GameLift create a StartMatchBackfill request whenever a game session has one or more open slots. Learn more about manual and automatic backfill in Backfill Existing Games with FlexMatch. Automatic backfill is not available when FlexMatchMode is set to STANDALONE.

createMatchmakingConfiguration_gameProperties :: Lens' CreateMatchmakingConfiguration (Maybe [GameProperty])

A set of custom properties for a game session, formatted as key:value pairs. These properties are passed to a game server process in the GameSession object with a request to start a new game session (see Start a Game Session). This information is added to the new GameSession object that is created for a successful match. This parameter is not used if FlexMatchMode is set to STANDALONE.

createMatchmakingConfiguration_acceptanceTimeoutSeconds :: Lens' CreateMatchmakingConfiguration (Maybe Natural)

The length of time (in seconds) to wait for players to accept a proposed match, if acceptance is required.

createMatchmakingConfiguration_additionalPlayerCount :: Lens' CreateMatchmakingConfiguration (Maybe Natural)

The number of player slots in a match to keep open for future players. For example, if the configuration's rule set specifies a match for a single 12-person team, and the additional player count is set to 2, only 10 players are selected for the match. This parameter is not used if FlexMatchMode is set to STANDALONE.

createMatchmakingConfiguration_gameSessionData :: Lens' CreateMatchmakingConfiguration (Maybe Text)

A set of custom game session properties, formatted as a single string value. This data is passed to a game server process in the GameSession object with a request to start a new game session (see Start a Game Session). This information is added to the new GameSession object that is created for a successful match. This parameter is not used if FlexMatchMode is set to STANDALONE.

createMatchmakingConfiguration_gameSessionQueueArns :: Lens' CreateMatchmakingConfiguration (Maybe [Text])

The Amazon Resource Name (ARN) that is assigned to a GameLift game session queue resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:<region>::gamesessionqueue/<queue name>. Queues can be located in any Region. Queues are used to start new GameLift-hosted game sessions for matches that are created with this matchmaking configuration. If FlexMatchMode is set to STANDALONE, do not set this parameter.

createMatchmakingConfiguration_notificationTarget :: Lens' CreateMatchmakingConfiguration (Maybe Text)

An SNS topic ARN that is set up to receive matchmaking notifications. See Setting up notifications for matchmaking for more information.

createMatchmakingConfiguration_tags :: Lens' CreateMatchmakingConfiguration (Maybe [Tag])

A list of labels to assign to the new matchmaking configuration resource. Tags are developer-defined key-value pairs. Tagging AWS resources are useful for resource management, access management and cost allocation. For more information, see Tagging AWS Resources in the AWS General Reference. Once the resource is created, you can use TagResource, UntagResource, and ListTagsForResource to add, remove, and view tags. The maximum tag limit may be lower than stated. See the AWS General Reference for actual tagging limits.

createMatchmakingConfiguration_description :: Lens' CreateMatchmakingConfiguration (Maybe Text)

A human-readable description of the matchmaking configuration.

createMatchmakingConfiguration_name :: Lens' CreateMatchmakingConfiguration Text

A unique identifier for the matchmaking configuration. This name is used to identify the configuration associated with a matchmaking request or ticket.

createMatchmakingConfiguration_requestTimeoutSeconds :: Lens' CreateMatchmakingConfiguration Natural

The maximum duration, in seconds, that a matchmaking ticket can remain in process before timing out. Requests that fail due to timing out can be resubmitted as needed.

createMatchmakingConfiguration_acceptanceRequired :: Lens' CreateMatchmakingConfiguration Bool

A flag that determines whether a match that was created with this configuration must be accepted by the matched players. To require acceptance, set to TRUE. With this option enabled, matchmaking tickets use the status REQUIRES_ACCEPTANCE to indicate when a completed potential match is waiting for player acceptance.

createMatchmakingConfiguration_ruleSetName :: Lens' CreateMatchmakingConfiguration Text

A unique identifier for the matchmaking rule set to use with this configuration. You can use either the rule set name or ARN value. A matchmaking configuration can only use rule sets that are defined in the same Region.

Destructuring the Response

data CreateMatchmakingConfigurationResponse

Represents the returned data in response to a request operation.

See: newCreateMatchmakingConfigurationResponse smart constructor.

Constructors

CreateMatchmakingConfigurationResponse'

Fields

configuration :: Maybe MatchmakingConfiguration

Object that describes the newly created matchmaking configuration.

httpStatus :: Int

The response's http status code.

newCreateMatchmakingConfigurationResponse

Arguments

:: Int
$sel:httpStatus:CreateMatchmakingConfigurationResponse'

-> CreateMatchmakingConfigurationResponse

Create a value of CreateMatchmakingConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configuration:CreateMatchmakingConfigurationResponse', createMatchmakingConfigurationResponse_configuration - Object that describes the newly created matchmaking configuration.

$sel:httpStatus:CreateMatchmakingConfigurationResponse', createMatchmakingConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

createMatchmakingConfigurationResponse_configuration :: Lens' CreateMatchmakingConfigurationResponse (Maybe MatchmakingConfiguration)

Object that describes the newly created matchmaking configuration.

createMatchmakingConfigurationResponse_httpStatus :: Lens' CreateMatchmakingConfigurationResponse Int

The response's http status code.