Network.AWS.GameLift.CreateMatchmakingRuleSet

Description

Creates a new rule set for FlexMatch matchmaking. A rule set describes the type of match to create, such as the number and size of teams. It also sets the parameters for acceptable player matches, such as minimum skill level or character type. A rule set is used by a MatchmakingConfiguration.

To create a matchmaking rule set, provide unique rule set name and the rule set body in JSON format. Rule sets must be defined in the same Region as the matchmaking configuration they are used with.

Since matchmaking rule sets cannot be edited, it is a good idea to check the rule set syntax using ValidateMatchmakingRuleSet before creating a new rule set.

Learn more

Build a rule set

Design a matchmaker

Matchmaking with FlexMatch

Related actions

CreateMatchmakingConfiguration | DescribeMatchmakingConfigurations | UpdateMatchmakingConfiguration | DeleteMatchmakingConfiguration | CreateMatchmakingRuleSet | DescribeMatchmakingRuleSets | ValidateMatchmakingRuleSet | DeleteMatchmakingRuleSet | All APIs by task

Creating a Request

data CreateMatchmakingRuleSet

Represents the input for a request operation.

See: newCreateMatchmakingRuleSet smart constructor.

Constructors

CreateMatchmakingRuleSet'

Fields

tags :: Maybe [Tag]

A list of labels to assign to the new matchmaking rule set resource. Tags are developer-defined key-value pairs. Tagging AWS resources are useful for resource management, access management and cost allocation. For more information, see Tagging AWS Resources in the AWS General Reference. Once the resource is created, you can use TagResource, UntagResource, and ListTagsForResource to add, remove, and view tags. The maximum tag limit may be lower than stated. See the AWS General Reference for actual tagging limits.

name :: Text

A unique identifier for the matchmaking rule set. A matchmaking configuration identifies the rule set it uses by this name value. Note that the rule set name is different from the optional name field in the rule set body.

ruleSetBody :: Text

A collection of matchmaking rules, formatted as a JSON string. Comments are not allowed in JSON, but most elements support a description field.

newCreateMatchmakingRuleSet

Arguments

:: Text
$sel:name:CreateMatchmakingRuleSet'

-> Text
$sel:ruleSetBody:CreateMatchmakingRuleSet'

-> CreateMatchmakingRuleSet

Create a value of CreateMatchmakingRuleSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateMatchmakingRuleSet', createMatchmakingRuleSet_tags - A list of labels to assign to the new matchmaking rule set resource. Tags are developer-defined key-value pairs. Tagging AWS resources are useful for resource management, access management and cost allocation. For more information, see Tagging AWS Resources in the AWS General Reference. Once the resource is created, you can use TagResource, UntagResource, and ListTagsForResource to add, remove, and view tags. The maximum tag limit may be lower than stated. See the AWS General Reference for actual tagging limits.

$sel:name:CreateMatchmakingRuleSet', createMatchmakingRuleSet_name - A unique identifier for the matchmaking rule set. A matchmaking configuration identifies the rule set it uses by this name value. Note that the rule set name is different from the optional name field in the rule set body.

$sel:ruleSetBody:CreateMatchmakingRuleSet', createMatchmakingRuleSet_ruleSetBody - A collection of matchmaking rules, formatted as a JSON string. Comments are not allowed in JSON, but most elements support a description field.

Request Lenses

createMatchmakingRuleSet_tags :: Lens' CreateMatchmakingRuleSet (Maybe [Tag])

A list of labels to assign to the new matchmaking rule set resource. Tags are developer-defined key-value pairs. Tagging AWS resources are useful for resource management, access management and cost allocation. For more information, see Tagging AWS Resources in the AWS General Reference. Once the resource is created, you can use TagResource, UntagResource, and ListTagsForResource to add, remove, and view tags. The maximum tag limit may be lower than stated. See the AWS General Reference for actual tagging limits.

createMatchmakingRuleSet_name :: Lens' CreateMatchmakingRuleSet Text

A unique identifier for the matchmaking rule set. A matchmaking configuration identifies the rule set it uses by this name value. Note that the rule set name is different from the optional name field in the rule set body.

createMatchmakingRuleSet_ruleSetBody :: Lens' CreateMatchmakingRuleSet Text

A collection of matchmaking rules, formatted as a JSON string. Comments are not allowed in JSON, but most elements support a description field.

Destructuring the Response

data CreateMatchmakingRuleSetResponse

Represents the returned data in response to a request operation.

See: newCreateMatchmakingRuleSetResponse smart constructor.

Constructors

CreateMatchmakingRuleSetResponse'

Fields

httpStatus :: Int

The response's http status code.

ruleSet :: MatchmakingRuleSet

The newly created matchmaking rule set.

newCreateMatchmakingRuleSetResponse

Arguments

:: Int
$sel:httpStatus:CreateMatchmakingRuleSetResponse'

-> MatchmakingRuleSet
$sel:ruleSet:CreateMatchmakingRuleSetResponse'

-> CreateMatchmakingRuleSetResponse

Create a value of CreateMatchmakingRuleSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateMatchmakingRuleSetResponse', createMatchmakingRuleSetResponse_httpStatus - The response's http status code.

$sel:ruleSet:CreateMatchmakingRuleSetResponse', createMatchmakingRuleSetResponse_ruleSet - The newly created matchmaking rule set.

Response Lenses

createMatchmakingRuleSetResponse_httpStatus :: Lens' CreateMatchmakingRuleSetResponse Int

The response's http status code.

createMatchmakingRuleSetResponse_ruleSet :: Lens' CreateMatchmakingRuleSetResponse MatchmakingRuleSet

The newly created matchmaking rule set.