Network.AWS.GameLift.CreateVpcPeeringConnection

Description

Establishes a VPC peering connection between a virtual private cloud (VPC) in an AWS account with the VPC for your Amazon GameLift fleet. VPC peering enables the game servers on your fleet to communicate directly with other AWS resources. You can peer with VPCs in any AWS account that you have access to, including the account that you use to manage your Amazon GameLift fleets. You cannot peer with VPCs that are in different Regions. For more information, see VPC Peering with Amazon GameLift Fleets.

Before calling this operation to establish the peering connection, you first need to call CreateVpcPeeringAuthorization and identify the VPC you want to peer with. Once the authorization for the specified VPC is issued, you have 24 hours to establish the connection. These two operations handle all tasks necessary to peer the two VPCs, including acceptance, updating routing tables, etc.

To establish the connection, call this operation from the AWS account that is used to manage the Amazon GameLift fleets. Identify the following values: (1) The ID of the fleet you want to be enable a VPC peering connection for; (2) The AWS account with the VPC that you want to peer with; and (3) The ID of the VPC you want to peer with. This operation is asynchronous. If successful, a VpcPeeringConnection request is created. You can use continuous polling to track the request's status using DescribeVpcPeeringConnections, or by monitoring fleet events for success or failure using DescribeFleetEvents.

Related actions

CreateVpcPeeringAuthorization | DescribeVpcPeeringAuthorizations | DeleteVpcPeeringAuthorization | CreateVpcPeeringConnection | DescribeVpcPeeringConnections | DeleteVpcPeeringConnection | All APIs by task

Creating a Request

data CreateVpcPeeringConnection

Represents the input for a request operation.

See: newCreateVpcPeeringConnection smart constructor.

Constructors

CreateVpcPeeringConnection'

Fields

fleetId :: Text

A unique identifier for the fleet. You can use either the fleet ID or ARN value. This tells Amazon GameLift which GameLift VPC to peer with.

peerVpcAwsAccountId :: Text

A unique identifier for the AWS account with the VPC that you want to peer your Amazon GameLift fleet with. You can find your Account ID in the AWS Management Console under account settings.

peerVpcId :: Text

A unique identifier for a VPC with resources to be accessed by your GameLift fleet. The VPC must be in the same Region as your fleet. To look up a VPC ID, use the VPC Dashboard in the AWS Management Console. Learn more about VPC peering in VPC Peering with GameLift Fleets.

newCreateVpcPeeringConnection

Arguments

:: Text
$sel:fleetId:CreateVpcPeeringConnection'

-> Text
$sel:peerVpcAwsAccountId:CreateVpcPeeringConnection'

-> Text
$sel:peerVpcId:CreateVpcPeeringConnection'

-> CreateVpcPeeringConnection

Create a value of CreateVpcPeeringConnection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fleetId:CreateVpcPeeringConnection', createVpcPeeringConnection_fleetId - A unique identifier for the fleet. You can use either the fleet ID or ARN value. This tells Amazon GameLift which GameLift VPC to peer with.

$sel:peerVpcAwsAccountId:CreateVpcPeeringConnection', createVpcPeeringConnection_peerVpcAwsAccountId - A unique identifier for the AWS account with the VPC that you want to peer your Amazon GameLift fleet with. You can find your Account ID in the AWS Management Console under account settings.

$sel:peerVpcId:CreateVpcPeeringConnection', createVpcPeeringConnection_peerVpcId - A unique identifier for a VPC with resources to be accessed by your GameLift fleet. The VPC must be in the same Region as your fleet. To look up a VPC ID, use the VPC Dashboard in the AWS Management Console. Learn more about VPC peering in VPC Peering with GameLift Fleets.

Request Lenses

createVpcPeeringConnection_fleetId :: Lens' CreateVpcPeeringConnection Text

A unique identifier for the fleet. You can use either the fleet ID or ARN value. This tells Amazon GameLift which GameLift VPC to peer with.

createVpcPeeringConnection_peerVpcAwsAccountId :: Lens' CreateVpcPeeringConnection Text

A unique identifier for the AWS account with the VPC that you want to peer your Amazon GameLift fleet with. You can find your Account ID in the AWS Management Console under account settings.

createVpcPeeringConnection_peerVpcId :: Lens' CreateVpcPeeringConnection Text

A unique identifier for a VPC with resources to be accessed by your GameLift fleet. The VPC must be in the same Region as your fleet. To look up a VPC ID, use the VPC Dashboard in the AWS Management Console. Learn more about VPC peering in VPC Peering with GameLift Fleets.

Destructuring the Response

data CreateVpcPeeringConnectionResponse

See: newCreateVpcPeeringConnectionResponse smart constructor.

Constructors

CreateVpcPeeringConnectionResponse'

Fields

httpStatus :: Int

The response's http status code.

newCreateVpcPeeringConnectionResponse

Arguments

:: Int
$sel:httpStatus:CreateVpcPeeringConnectionResponse'

-> CreateVpcPeeringConnectionResponse

Create a value of CreateVpcPeeringConnectionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateVpcPeeringConnectionResponse', createVpcPeeringConnectionResponse_httpStatus - The response's http status code.

Response Lenses

createVpcPeeringConnectionResponse_httpStatus :: Lens' CreateVpcPeeringConnectionResponse Int

The response's http status code.