Network.AWS.GameLift.DeleteFleetLocations

Description

Removes locations from a multi-location fleet. When deleting a location, all game server process and all instances that are still active in the location are shut down.

To delete fleet locations, identify the fleet ID and provide a list of the locations to be deleted.

If successful, GameLift sets the location status to DELETING, and begins to shut down existing server processes and terminate instances in each location being deleted. When completed, the location status changes to TERMINATED.

Learn more

Setting up GameLift fleets

Related actions

CreateFleetLocations | DescribeFleetLocationAttributes | DescribeFleetLocationCapacity | DescribeFleetLocationUtilization | DescribeFleetAttributes | DescribeFleetCapacity | DescribeFleetUtilization | UpdateFleetCapacity | StopFleetActions | DeleteFleetLocations | All APIs by task

Creating a Request

data DeleteFleetLocations

Represents the input for a request operation.

See: newDeleteFleetLocations smart constructor.

Constructors

DeleteFleetLocations'

Fields

fleetId :: Text

A unique identifier for the fleet to delete locations for. You can use either the fleet ID or ARN value.

locations :: NonEmpty Text

The list of fleet locations to delete. Specify locations in the form of an AWS Region code, such as us-west-2.

newDeleteFleetLocations

Arguments

:: Text
$sel:fleetId:DeleteFleetLocations'

-> NonEmpty Text
$sel:locations:DeleteFleetLocations'

-> DeleteFleetLocations

Create a value of DeleteFleetLocations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fleetId:DeleteFleetLocations', deleteFleetLocations_fleetId - A unique identifier for the fleet to delete locations for. You can use either the fleet ID or ARN value.

$sel:locations:DeleteFleetLocations', deleteFleetLocations_locations - The list of fleet locations to delete. Specify locations in the form of an AWS Region code, such as us-west-2.

Request Lenses

deleteFleetLocations_fleetId :: Lens' DeleteFleetLocations Text

A unique identifier for the fleet to delete locations for. You can use either the fleet ID or ARN value.

deleteFleetLocations_locations :: Lens' DeleteFleetLocations (NonEmpty Text)

The list of fleet locations to delete. Specify locations in the form of an AWS Region code, such as us-west-2.

Destructuring the Response

data DeleteFleetLocationsResponse

Represents the returned data in response to a request operation.

See: newDeleteFleetLocationsResponse smart constructor.

Constructors

DeleteFleetLocationsResponse'

Fields

fleetArn :: Maybe Text

The Amazon Resource Name (ARN) that is assigned to a GameLift fleet resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:<region>::fleet/fleet-a1234567-b8c9-0d1e-2fa3-b45c6d7e8912.

locationStates :: Maybe [LocationState]

The remote locations that are being deleted, with each location status set to DELETING.

fleetId :: Maybe Text

A unique identifier for the fleet that location attributes are being deleted for.

httpStatus :: Int

The response's http status code.

newDeleteFleetLocationsResponse

Arguments

:: Int
$sel:httpStatus:DeleteFleetLocationsResponse'

-> DeleteFleetLocationsResponse

Create a value of DeleteFleetLocationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fleetArn:DeleteFleetLocationsResponse', deleteFleetLocationsResponse_fleetArn - The Amazon Resource Name (ARN) that is assigned to a GameLift fleet resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:<region>::fleet/fleet-a1234567-b8c9-0d1e-2fa3-b45c6d7e8912.

$sel:locationStates:DeleteFleetLocationsResponse', deleteFleetLocationsResponse_locationStates - The remote locations that are being deleted, with each location status set to DELETING.

$sel:fleetId:DeleteFleetLocations', deleteFleetLocationsResponse_fleetId - A unique identifier for the fleet that location attributes are being deleted for.

$sel:httpStatus:DeleteFleetLocationsResponse', deleteFleetLocationsResponse_httpStatus - The response's http status code.

Response Lenses

deleteFleetLocationsResponse_fleetArn :: Lens' DeleteFleetLocationsResponse (Maybe Text)

The Amazon Resource Name (ARN) that is assigned to a GameLift fleet resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:<region>::fleet/fleet-a1234567-b8c9-0d1e-2fa3-b45c6d7e8912.

deleteFleetLocationsResponse_locationStates :: Lens' DeleteFleetLocationsResponse (Maybe [LocationState])

The remote locations that are being deleted, with each location status set to DELETING.

deleteFleetLocationsResponse_fleetId :: Lens' DeleteFleetLocationsResponse (Maybe Text)

A unique identifier for the fleet that location attributes are being deleted for.

deleteFleetLocationsResponse_httpStatus :: Lens' DeleteFleetLocationsResponse Int

The response's http status code.