Network.AWS.GameLift.DeleteGameServerGroup

Description

__This operation is used with the GameLift FleetIQ solution and game server groups.__

Terminates a game server group and permanently deletes the game server group record. You have several options for how these resources are impacted when deleting the game server group. Depending on the type of delete operation selected, this operation might affect these resources:

The game server group

The corresponding Auto Scaling group

All game servers that are currently running in the group

To delete a game server group, identify the game server group to delete and specify the type of delete operation to initiate. Game server groups can only be deleted if they are in ACTIVE or ERROR status.

If the delete request is successful, a series of operations are kicked off. The game server group status is changed to DELETE_SCHEDULED, which prevents new game servers from being registered and stops automatic scaling activity. Once all game servers in the game server group are deregistered, GameLift FleetIQ can begin deleting resources. If any of the delete operations fail, the game server group is placed in ERROR status.

GameLift FleetIQ emits delete events to Amazon CloudWatch.

Learn more

GameLift FleetIQ Guide

Related actions

CreateGameServerGroup | ListGameServerGroups | DescribeGameServerGroup | UpdateGameServerGroup | DeleteGameServerGroup | ResumeGameServerGroup | SuspendGameServerGroup | DescribeGameServerInstances | All APIs by task

Creating a Request

data DeleteGameServerGroup

See: newDeleteGameServerGroup smart constructor.

Constructors

DeleteGameServerGroup'

Fields

deleteOption :: Maybe GameServerGroupDeleteOption

The type of delete to perform. Options include the following:

SAFE_DELETE – (default) Terminates the game server group and EC2 Auto Scaling group only when it has no game servers that are in UTILIZED status.

FORCE_DELETE – Terminates the game server group, including all active game servers regardless of their utilization status, and the EC2 Auto Scaling group.

RETAIN – Does a safe delete of the game server group but retains the EC2 Auto Scaling group as is.

gameServerGroupName :: Text

A unique identifier for the game server group. Use either the GameServerGroup name or ARN value.

newDeleteGameServerGroup

Arguments

:: Text
$sel:gameServerGroupName:DeleteGameServerGroup'

-> DeleteGameServerGroup

Create a value of DeleteGameServerGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deleteOption:DeleteGameServerGroup', deleteGameServerGroup_deleteOption - The type of delete to perform. Options include the following:

SAFE_DELETE – (default) Terminates the game server group and EC2 Auto Scaling group only when it has no game servers that are in UTILIZED status.

FORCE_DELETE – Terminates the game server group, including all active game servers regardless of their utilization status, and the EC2 Auto Scaling group.

RETAIN – Does a safe delete of the game server group but retains the EC2 Auto Scaling group as is.

$sel:gameServerGroupName:DeleteGameServerGroup', deleteGameServerGroup_gameServerGroupName - A unique identifier for the game server group. Use either the GameServerGroup name or ARN value.

Request Lenses

deleteGameServerGroup_deleteOption :: Lens' DeleteGameServerGroup (Maybe GameServerGroupDeleteOption)

The type of delete to perform. Options include the following:

SAFE_DELETE – (default) Terminates the game server group and EC2 Auto Scaling group only when it has no game servers that are in UTILIZED status.

FORCE_DELETE – Terminates the game server group, including all active game servers regardless of their utilization status, and the EC2 Auto Scaling group.

RETAIN – Does a safe delete of the game server group but retains the EC2 Auto Scaling group as is.

deleteGameServerGroup_gameServerGroupName :: Lens' DeleteGameServerGroup Text

A unique identifier for the game server group. Use either the GameServerGroup name or ARN value.

Destructuring the Response

data DeleteGameServerGroupResponse

See: newDeleteGameServerGroupResponse smart constructor.

Constructors

DeleteGameServerGroupResponse'

Fields

gameServerGroup :: Maybe GameServerGroup

An object that describes the deleted game server group resource, with status updated to DELETE_SCHEDULED.

httpStatus :: Int

The response's http status code.

newDeleteGameServerGroupResponse

Arguments

:: Int
$sel:httpStatus:DeleteGameServerGroupResponse'

-> DeleteGameServerGroupResponse

Create a value of DeleteGameServerGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:gameServerGroup:DeleteGameServerGroupResponse', deleteGameServerGroupResponse_gameServerGroup - An object that describes the deleted game server group resource, with status updated to DELETE_SCHEDULED.

$sel:httpStatus:DeleteGameServerGroupResponse', deleteGameServerGroupResponse_httpStatus - The response's http status code.

Response Lenses

deleteGameServerGroupResponse_gameServerGroup :: Lens' DeleteGameServerGroupResponse (Maybe GameServerGroup)

An object that describes the deleted game server group resource, with status updated to DELETE_SCHEDULED.

deleteGameServerGroupResponse_httpStatus :: Lens' DeleteGameServerGroupResponse Int

The response's http status code.