Network.AWS.GameLift.DeleteVpcPeeringConnection

Description

Removes a VPC peering connection. To delete the connection, you must have a valid authorization for the VPC peering connection that you want to delete. You can check for an authorization by calling DescribeVpcPeeringAuthorizations or request a new one using CreateVpcPeeringAuthorization.

Once a valid authorization exists, call this operation from the AWS account that is used to manage the Amazon GameLift fleets. Identify the connection to delete by the connection ID and fleet ID. If successful, the connection is removed.

Related actions

CreateVpcPeeringAuthorization | DescribeVpcPeeringAuthorizations | DeleteVpcPeeringAuthorization | CreateVpcPeeringConnection | DescribeVpcPeeringConnections | DeleteVpcPeeringConnection | All APIs by task

Creating a Request

data DeleteVpcPeeringConnection

Represents the input for a request operation.

See: newDeleteVpcPeeringConnection smart constructor.

Constructors

DeleteVpcPeeringConnection'

Fields

fleetId :: Text

A unique identifier for the fleet. This fleet specified must match the fleet referenced in the VPC peering connection record. You can use either the fleet ID or ARN value.

vpcPeeringConnectionId :: Text

A unique identifier for a VPC peering connection. This value is included in the VpcPeeringConnection object, which can be retrieved by calling DescribeVpcPeeringConnections.

newDeleteVpcPeeringConnection

Arguments

:: Text
$sel:fleetId:DeleteVpcPeeringConnection'

-> Text
$sel:vpcPeeringConnectionId:DeleteVpcPeeringConnection'

-> DeleteVpcPeeringConnection

Create a value of DeleteVpcPeeringConnection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fleetId:DeleteVpcPeeringConnection', deleteVpcPeeringConnection_fleetId - A unique identifier for the fleet. This fleet specified must match the fleet referenced in the VPC peering connection record. You can use either the fleet ID or ARN value.

$sel:vpcPeeringConnectionId:DeleteVpcPeeringConnection', deleteVpcPeeringConnection_vpcPeeringConnectionId - A unique identifier for a VPC peering connection. This value is included in the VpcPeeringConnection object, which can be retrieved by calling DescribeVpcPeeringConnections.

Request Lenses

deleteVpcPeeringConnection_fleetId :: Lens' DeleteVpcPeeringConnection Text

A unique identifier for the fleet. This fleet specified must match the fleet referenced in the VPC peering connection record. You can use either the fleet ID or ARN value.

deleteVpcPeeringConnection_vpcPeeringConnectionId :: Lens' DeleteVpcPeeringConnection Text

A unique identifier for a VPC peering connection. This value is included in the VpcPeeringConnection object, which can be retrieved by calling DescribeVpcPeeringConnections.

Destructuring the Response

data DeleteVpcPeeringConnectionResponse

See: newDeleteVpcPeeringConnectionResponse smart constructor.

Constructors

DeleteVpcPeeringConnectionResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteVpcPeeringConnectionResponse

Arguments

:: Int
$sel:httpStatus:DeleteVpcPeeringConnectionResponse'

-> DeleteVpcPeeringConnectionResponse

Create a value of DeleteVpcPeeringConnectionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteVpcPeeringConnectionResponse', deleteVpcPeeringConnectionResponse_httpStatus - The response's http status code.

Response Lenses

deleteVpcPeeringConnectionResponse_httpStatus :: Lens' DeleteVpcPeeringConnectionResponse Int

The response's http status code.