Network.AWS.GameLift.DescribeEC2InstanceLimits

Description

The GameLift service limits and current utilization for an AWS Region or location. Instance limits control the number of instances, per instance type, per location, that your AWS account can use. Learn more at Amazon EC2 Instance Types. The information returned includes the maximum number of instances allowed and your account's current usage across all fleets. This information can affect your ability to scale your GameLift fleets. You can request a limit increase for your account by using the __Service limits__ page in the GameLift console.

Instance limits differ based on whether the instances are deployed in a fleet's home Region or in a remote location. For remote locations, limits also differ based on the combination of home Region and remote location. All requests must specify an AWS Region (either explicitly or as your default settings). To get the limit for a remote location, you must also specify the location. For example, the following requests all return different results:

Request specifies the Region ap-northeast-1 with no location. The result is limits and usage data on all instance types that are deployed in us-east-2, by all of the fleets that reside in ap-northeast-1.

Request specifies the Region us-east-1 with location ca-central-1. The result is limits and usage data on all instance types that are deployed in ca-central-1, by all of the fleets that reside in us-east-2. These limits do not affect fleets in any other Regions that deploy instances to ca-central-1.

Request specifies the Region eu-west-1 with location ca-central-1. The result is limits and usage data on all instance types that are deployed in ca-central-1, by all of the fleets that reside in eu-west-1.

This operation can be used in the following ways:

To get limit and usage data for all instance types that are deployed in an AWS Region by fleets that reside in the same Region: Specify the Region only. Optionally, specify a single instance type to retrieve information for.

To get limit and usage data for all instance types that are deployed to a remote location by fleets that reside in different AWS Region: Provide both the AWS Region and the remote location. Optionally, specify a single instance type to retrieve information for.

If successful, an EC2InstanceLimits object is returned with limits and usage data for each requested instance type.

Learn more

Setting up GameLift fleets

Related actions

CreateFleet | UpdateFleetCapacity | PutScalingPolicy | DescribeEC2InstanceLimits | DescribeFleetAttributes | DescribeFleetLocationAttributes | UpdateFleetAttributes | StopFleetActions | DeleteFleet | All APIs by task

Creating a Request

data DescribeEC2InstanceLimits

Represents the input for a request operation.

See: newDescribeEC2InstanceLimits smart constructor.

Constructors

DescribeEC2InstanceLimits'

Fields

location :: Maybe Text

The name of a remote location to request instance limits for, in the form of an AWS Region code such as us-west-2.

eC2InstanceType :: Maybe EC2InstanceType

Name of an EC2 instance type that is supported in GameLift. A fleet instance type determines the computing resources of each instance in the fleet, including CPU, memory, storage, and networking capacity. Do not specify a value for this parameter to retrieve limits for all instance types.

newDescribeEC2InstanceLimits :: DescribeEC2InstanceLimits

Create a value of DescribeEC2InstanceLimits with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:location:DescribeEC2InstanceLimits', describeEC2InstanceLimits_location - The name of a remote location to request instance limits for, in the form of an AWS Region code such as us-west-2.

$sel:eC2InstanceType:DescribeEC2InstanceLimits', describeEC2InstanceLimits_eC2InstanceType - Name of an EC2 instance type that is supported in GameLift. A fleet instance type determines the computing resources of each instance in the fleet, including CPU, memory, storage, and networking capacity. Do not specify a value for this parameter to retrieve limits for all instance types.

Request Lenses

describeEC2InstanceLimits_location :: Lens' DescribeEC2InstanceLimits (Maybe Text)

The name of a remote location to request instance limits for, in the form of an AWS Region code such as us-west-2.

describeEC2InstanceLimits_eC2InstanceType :: Lens' DescribeEC2InstanceLimits (Maybe EC2InstanceType)

Name of an EC2 instance type that is supported in GameLift. A fleet instance type determines the computing resources of each instance in the fleet, including CPU, memory, storage, and networking capacity. Do not specify a value for this parameter to retrieve limits for all instance types.

Destructuring the Response

data DescribeEC2InstanceLimitsResponse

Represents the returned data in response to a request operation.

See: newDescribeEC2InstanceLimitsResponse smart constructor.

Constructors

DescribeEC2InstanceLimitsResponse'

Fields

eC2InstanceLimits :: Maybe [EC2InstanceLimit]

The maximum number of instances for the specified instance type.

httpStatus :: Int

The response's http status code.

newDescribeEC2InstanceLimitsResponse

Arguments

:: Int
$sel:httpStatus:DescribeEC2InstanceLimitsResponse'

-> DescribeEC2InstanceLimitsResponse

Create a value of DescribeEC2InstanceLimitsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eC2InstanceLimits:DescribeEC2InstanceLimitsResponse', describeEC2InstanceLimitsResponse_eC2InstanceLimits - The maximum number of instances for the specified instance type.

$sel:httpStatus:DescribeEC2InstanceLimitsResponse', describeEC2InstanceLimitsResponse_httpStatus - The response's http status code.

Response Lenses

describeEC2InstanceLimitsResponse_eC2InstanceLimits :: Lens' DescribeEC2InstanceLimitsResponse (Maybe [EC2InstanceLimit])

The maximum number of instances for the specified instance type.

describeEC2InstanceLimitsResponse_httpStatus :: Lens' DescribeEC2InstanceLimitsResponse Int

The response's http status code.