Network.AWS.GameLift.DescribeFleetCapacity

Description

Retrieves the resource capacity settings for one or more fleets. The data returned includes the current fleet capacity (number of EC2 instances), and settings that can control how capacity scaling. For fleets with remote locations, this operation retrieves data for the fleet's home Region only. See DescribeFleetLocationCapacity to get capacity settings for a fleet's remote locations.

This operation can be used in the following ways:

To get capacity data for one or more specific fleets, provide a list of fleet IDs or fleet ARNs.

To get capacity data for all fleets, do not provide a fleet identifier.

When requesting multiple fleets, use the pagination parameters to retrieve results as a set of sequential pages.

If successful, a FleetCapacity object is returned for each requested fleet ID. Each FleetCapacity object includes a Location property, which is set to the fleet's home Region. When a list of fleet IDs is provided, attribute objects are returned only for fleets that currently exist.

Some API operations may limit the number of fleet IDs that are allowed in one request. If a request exceeds this limit, the request fails and the error message includes the maximum allowed.

Learn more

Setting up GameLift fleets

GameLift metrics for fleets

Related actions

ListFleets | DescribeEC2InstanceLimits | DescribeFleetAttributes | DescribeFleetCapacity | DescribeFleetEvents | DescribeFleetLocationAttributes | DescribeFleetPortSettings | DescribeFleetUtilization | DescribeRuntimeConfiguration | DescribeScalingPolicies | All APIs by task

This operation returns paginated results.

Creating a Request

data DescribeFleetCapacity

Represents the input for a request operation.

See: newDescribeFleetCapacity smart constructor.

Constructors

DescribeFleetCapacity'

Fields

nextToken :: Maybe Text

A token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this operation. To start at the beginning of the result set, do not specify a value. This parameter is ignored when the request specifies one or a list of fleet IDs.

fleetIds :: Maybe (NonEmpty Text)

A unique identifier for the fleet(s) to retrieve capacity information for. You can use either the fleet ID or ARN value. Leave this parameter empty to retrieve capacity information for all fleets.

limit :: Maybe Natural

The maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages. This parameter is ignored when the request specifies one or a list of fleet IDs.

newDescribeFleetCapacity :: DescribeFleetCapacity

Create a value of DescribeFleetCapacity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeFleetCapacity', describeFleetCapacity_nextToken - A token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this operation. To start at the beginning of the result set, do not specify a value. This parameter is ignored when the request specifies one or a list of fleet IDs.

$sel:fleetIds:DescribeFleetCapacity', describeFleetCapacity_fleetIds - A unique identifier for the fleet(s) to retrieve capacity information for. You can use either the fleet ID or ARN value. Leave this parameter empty to retrieve capacity information for all fleets.

$sel:limit:DescribeFleetCapacity', describeFleetCapacity_limit - The maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages. This parameter is ignored when the request specifies one or a list of fleet IDs.

Request Lenses

describeFleetCapacity_nextToken :: Lens' DescribeFleetCapacity (Maybe Text)

A token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this operation. To start at the beginning of the result set, do not specify a value. This parameter is ignored when the request specifies one or a list of fleet IDs.

describeFleetCapacity_fleetIds :: Lens' DescribeFleetCapacity (Maybe (NonEmpty Text))

A unique identifier for the fleet(s) to retrieve capacity information for. You can use either the fleet ID or ARN value. Leave this parameter empty to retrieve capacity information for all fleets.

describeFleetCapacity_limit :: Lens' DescribeFleetCapacity (Maybe Natural)

The maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages. This parameter is ignored when the request specifies one or a list of fleet IDs.

Destructuring the Response

data DescribeFleetCapacityResponse

Represents the returned data in response to a request operation.

See: newDescribeFleetCapacityResponse smart constructor.

Constructors

DescribeFleetCapacityResponse'

Fields

nextToken :: Maybe Text

A token that indicates where to resume retrieving results on the next call to this operation. If no token is returned, these results represent the end of the list.

fleetCapacity :: Maybe [FleetCapacity]

A collection of objects that contains capacity information for each requested fleet ID. Capacity objects are returned only for fleets that currently exist.

httpStatus :: Int

The response's http status code.

newDescribeFleetCapacityResponse

Arguments

:: Int
$sel:httpStatus:DescribeFleetCapacityResponse'

-> DescribeFleetCapacityResponse

Create a value of DescribeFleetCapacityResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeFleetCapacity', describeFleetCapacityResponse_nextToken - A token that indicates where to resume retrieving results on the next call to this operation. If no token is returned, these results represent the end of the list.

$sel:fleetCapacity:DescribeFleetCapacityResponse', describeFleetCapacityResponse_fleetCapacity - A collection of objects that contains capacity information for each requested fleet ID. Capacity objects are returned only for fleets that currently exist.

$sel:httpStatus:DescribeFleetCapacityResponse', describeFleetCapacityResponse_httpStatus - The response's http status code.

Response Lenses

describeFleetCapacityResponse_nextToken :: Lens' DescribeFleetCapacityResponse (Maybe Text)

A token that indicates where to resume retrieving results on the next call to this operation. If no token is returned, these results represent the end of the list.

describeFleetCapacityResponse_fleetCapacity :: Lens' DescribeFleetCapacityResponse (Maybe [FleetCapacity])

A collection of objects that contains capacity information for each requested fleet ID. Capacity objects are returned only for fleets that currently exist.

describeFleetCapacityResponse_httpStatus :: Lens' DescribeFleetCapacityResponse Int

The response's http status code.