Network.AWS.GameLift.DescribeFleetLocationAttributes

Description

Retrieves information on a fleet's remote locations, including life-cycle status and any suspended fleet activity.

This operation can be used in the following ways:

To get data for specific locations, provide a fleet identifier and a list of locations. Location data is returned in the order that it is requested.

To get data for all locations, provide a fleet identifier only. Location data is returned in no particular order.

When requesting attributes for multiple locations, use the pagination parameters to retrieve results as a set of sequential pages.

If successful, a LocationAttributes object is returned for each requested location. If the fleet does not have a requested location, no information is returned. This operation does not return the home Region. To get information on a fleet's home Region, call DescribeFleetAttributes.

Learn more

Setting up GameLift fleets

Related actions

CreateFleetLocations | DescribeFleetLocationAttributes | DescribeFleetLocationCapacity | DescribeFleetLocationUtilization | DescribeFleetAttributes | DescribeFleetCapacity | DescribeFleetUtilization | UpdateFleetCapacity | StopFleetActions | DeleteFleetLocations | All APIs by task

Creating a Request

data DescribeFleetLocationAttributes

Represents the input for a request operation.

See: newDescribeFleetLocationAttributes smart constructor.

Constructors

DescribeFleetLocationAttributes'

Fields

nextToken :: Maybe Text

A token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this operation. To start at the beginning of the result set, do not specify a value.

locations :: Maybe (NonEmpty Text)

A list of fleet locations to retrieve information for. Specify locations in the form of an AWS Region code, such as us-west-2.

limit :: Maybe Natural

The maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages. This limit is not currently enforced.

fleetId :: Text

A unique identifier for the fleet to retrieve remote locations for. You can use either the fleet ID or ARN value.

newDescribeFleetLocationAttributes

Arguments

:: Text
$sel:fleetId:DescribeFleetLocationAttributes'

-> DescribeFleetLocationAttributes

Create a value of DescribeFleetLocationAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeFleetLocationAttributes', describeFleetLocationAttributes_nextToken - A token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this operation. To start at the beginning of the result set, do not specify a value.

$sel:locations:DescribeFleetLocationAttributes', describeFleetLocationAttributes_locations - A list of fleet locations to retrieve information for. Specify locations in the form of an AWS Region code, such as us-west-2.

$sel:limit:DescribeFleetLocationAttributes', describeFleetLocationAttributes_limit - The maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages. This limit is not currently enforced.

$sel:fleetId:DescribeFleetLocationAttributes', describeFleetLocationAttributes_fleetId - A unique identifier for the fleet to retrieve remote locations for. You can use either the fleet ID or ARN value.

Request Lenses

describeFleetLocationAttributes_nextToken :: Lens' DescribeFleetLocationAttributes (Maybe Text)

A token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this operation. To start at the beginning of the result set, do not specify a value.

describeFleetLocationAttributes_locations :: Lens' DescribeFleetLocationAttributes (Maybe (NonEmpty Text))

A list of fleet locations to retrieve information for. Specify locations in the form of an AWS Region code, such as us-west-2.

describeFleetLocationAttributes_limit :: Lens' DescribeFleetLocationAttributes (Maybe Natural)

The maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages. This limit is not currently enforced.

describeFleetLocationAttributes_fleetId :: Lens' DescribeFleetLocationAttributes Text

A unique identifier for the fleet to retrieve remote locations for. You can use either the fleet ID or ARN value.

Destructuring the Response

data DescribeFleetLocationAttributesResponse

Represents the returned data in response to a request operation.

See: newDescribeFleetLocationAttributesResponse smart constructor.

Constructors

DescribeFleetLocationAttributesResponse'

Fields

nextToken :: Maybe Text

A token that indicates where to resume retrieving results on the next call to this operation. If no token is returned, these results represent the end of the list.

fleetId :: Maybe Text

A unique identifier for the fleet that location attributes were requested for.

fleetArn :: Maybe Text

The Amazon Resource Name (ARN) that is assigned to a GameLift fleet resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:<region>::fleet/fleet-a1234567-b8c9-0d1e-2fa3-b45c6d7e8912.

locationAttributes :: Maybe [LocationAttributes]

Location-specific information on the requested fleet's remote locations.

httpStatus :: Int

The response's http status code.

newDescribeFleetLocationAttributesResponse

Arguments

:: Int
$sel:httpStatus:DescribeFleetLocationAttributesResponse'

-> DescribeFleetLocationAttributesResponse

Create a value of DescribeFleetLocationAttributesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeFleetLocationAttributes', describeFleetLocationAttributesResponse_nextToken - A token that indicates where to resume retrieving results on the next call to this operation. If no token is returned, these results represent the end of the list.

$sel:fleetId:DescribeFleetLocationAttributes', describeFleetLocationAttributesResponse_fleetId - A unique identifier for the fleet that location attributes were requested for.

$sel:fleetArn:DescribeFleetLocationAttributesResponse', describeFleetLocationAttributesResponse_fleetArn - The Amazon Resource Name (ARN) that is assigned to a GameLift fleet resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:<region>::fleet/fleet-a1234567-b8c9-0d1e-2fa3-b45c6d7e8912.

$sel:locationAttributes:DescribeFleetLocationAttributesResponse', describeFleetLocationAttributesResponse_locationAttributes - Location-specific information on the requested fleet's remote locations.

$sel:httpStatus:DescribeFleetLocationAttributesResponse', describeFleetLocationAttributesResponse_httpStatus - The response's http status code.

Response Lenses

describeFleetLocationAttributesResponse_nextToken :: Lens' DescribeFleetLocationAttributesResponse (Maybe Text)

A token that indicates where to resume retrieving results on the next call to this operation. If no token is returned, these results represent the end of the list.

describeFleetLocationAttributesResponse_fleetId :: Lens' DescribeFleetLocationAttributesResponse (Maybe Text)

A unique identifier for the fleet that location attributes were requested for.

describeFleetLocationAttributesResponse_fleetArn :: Lens' DescribeFleetLocationAttributesResponse (Maybe Text)

The Amazon Resource Name (ARN) that is assigned to a GameLift fleet resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:<region>::fleet/fleet-a1234567-b8c9-0d1e-2fa3-b45c6d7e8912.

describeFleetLocationAttributesResponse_locationAttributes :: Lens' DescribeFleetLocationAttributesResponse (Maybe [LocationAttributes])

Location-specific information on the requested fleet's remote locations.

describeFleetLocationAttributesResponse_httpStatus :: Lens' DescribeFleetLocationAttributesResponse Int

The response's http status code.