Network.AWS.GameLift.DescribeFleetUtilization

Description

Retrieves utilization statistics for one or more fleets. Utilization data provides a snapshot of how the fleet's hosting resources are currently being used. For fleets with remote locations, this operation retrieves data for the fleet's home Region only. See DescribeFleetLocationUtilization to get utilization statistics for a fleet's remote locations.

This operation can be used in the following ways:

To get utilization data for one or more specific fleets, provide a list of fleet IDs or fleet ARNs.

To get utilization data for all fleets, do not provide a fleet identifier.

When requesting multiple fleets, use the pagination parameters to retrieve results as a set of sequential pages.

If successful, a FleetUtilization object is returned for each requested fleet ID, unless the fleet identifier is not found. Each fleet utilization object includes a Location property, which is set to the fleet's home Region.

Some API operations may limit the number of fleet IDs allowed in one request. If a request exceeds this limit, the request fails and the error message includes the maximum allowed.

Learn more

Setting up GameLift Fleets

GameLift Metrics for Fleets

Related actions

ListFleets | DescribeEC2InstanceLimits | DescribeFleetAttributes | DescribeFleetCapacity | DescribeFleetEvents | DescribeFleetLocationAttributes | DescribeFleetPortSettings | DescribeFleetUtilization | DescribeRuntimeConfiguration | DescribeScalingPolicies | All APIs by task

This operation returns paginated results.

Creating a Request

data DescribeFleetUtilization

Represents the input for a request operation.

See: newDescribeFleetUtilization smart constructor.

Constructors

DescribeFleetUtilization'

Fields

nextToken :: Maybe Text

A token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this operation. To start at the beginning of the result set, do not specify a value. This parameter is ignored when the request specifies one or a list of fleet IDs.

fleetIds :: Maybe (NonEmpty Text)

A unique identifier for the fleet(s) to retrieve utilization data for. You can use either the fleet ID or ARN value. To retrieve attributes for all current fleets, do not include this parameter.

limit :: Maybe Natural

The maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages. This parameter is ignored when the request specifies one or a list of fleet IDs.

newDescribeFleetUtilization :: DescribeFleetUtilization

Create a value of DescribeFleetUtilization with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeFleetUtilization', describeFleetUtilization_nextToken - A token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this operation. To start at the beginning of the result set, do not specify a value. This parameter is ignored when the request specifies one or a list of fleet IDs.

$sel:fleetIds:DescribeFleetUtilization', describeFleetUtilization_fleetIds - A unique identifier for the fleet(s) to retrieve utilization data for. You can use either the fleet ID or ARN value. To retrieve attributes for all current fleets, do not include this parameter.

$sel:limit:DescribeFleetUtilization', describeFleetUtilization_limit - The maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages. This parameter is ignored when the request specifies one or a list of fleet IDs.

Request Lenses

describeFleetUtilization_nextToken :: Lens' DescribeFleetUtilization (Maybe Text)

A token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this operation. To start at the beginning of the result set, do not specify a value. This parameter is ignored when the request specifies one or a list of fleet IDs.

describeFleetUtilization_fleetIds :: Lens' DescribeFleetUtilization (Maybe (NonEmpty Text))

A unique identifier for the fleet(s) to retrieve utilization data for. You can use either the fleet ID or ARN value. To retrieve attributes for all current fleets, do not include this parameter.

describeFleetUtilization_limit :: Lens' DescribeFleetUtilization (Maybe Natural)

The maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages. This parameter is ignored when the request specifies one or a list of fleet IDs.

Destructuring the Response

data DescribeFleetUtilizationResponse

Represents the returned data in response to a request operation.

See: newDescribeFleetUtilizationResponse smart constructor.

Constructors

DescribeFleetUtilizationResponse'

Fields

nextToken :: Maybe Text

A token that indicates where to resume retrieving results on the next call to this operation. If no token is returned, these results represent the end of the list.

fleetUtilization :: Maybe [FleetUtilization]

A collection of objects containing utilization information for each requested fleet ID. Utilization objects are returned only for fleets that currently exist.

httpStatus :: Int

The response's http status code.

newDescribeFleetUtilizationResponse

Arguments

:: Int
$sel:httpStatus:DescribeFleetUtilizationResponse'

-> DescribeFleetUtilizationResponse

Create a value of DescribeFleetUtilizationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeFleetUtilization', describeFleetUtilizationResponse_nextToken - A token that indicates where to resume retrieving results on the next call to this operation. If no token is returned, these results represent the end of the list.

$sel:fleetUtilization:DescribeFleetUtilizationResponse', describeFleetUtilizationResponse_fleetUtilization - A collection of objects containing utilization information for each requested fleet ID. Utilization objects are returned only for fleets that currently exist.

$sel:httpStatus:DescribeFleetUtilizationResponse', describeFleetUtilizationResponse_httpStatus - The response's http status code.

Response Lenses

describeFleetUtilizationResponse_nextToken :: Lens' DescribeFleetUtilizationResponse (Maybe Text)

A token that indicates where to resume retrieving results on the next call to this operation. If no token is returned, these results represent the end of the list.

describeFleetUtilizationResponse_fleetUtilization :: Lens' DescribeFleetUtilizationResponse (Maybe [FleetUtilization])

A collection of objects containing utilization information for each requested fleet ID. Utilization objects are returned only for fleets that currently exist.

describeFleetUtilizationResponse_httpStatus :: Lens' DescribeFleetUtilizationResponse Int

The response's http status code.