Network.AWS.GameLift.DescribeGameServerGroup

Description

__This operation is used with the GameLift FleetIQ solution and game server groups.__

Retrieves information on a game server group. This operation returns only properties related to GameLift FleetIQ. To view or update properties for the corresponding Auto Scaling group, such as launch template, auto scaling policies, and maximum/minimum group size, access the Auto Scaling group directly.

To get attributes for a game server group, provide a group name or ARN value. If successful, a GameServerGroup object is returned.

Learn more

GameLift FleetIQ Guide

Related actions

CreateGameServerGroup | ListGameServerGroups | DescribeGameServerGroup | UpdateGameServerGroup | DeleteGameServerGroup | ResumeGameServerGroup | SuspendGameServerGroup | DescribeGameServerInstances | All APIs by task

Creating a Request

data DescribeGameServerGroup

See: newDescribeGameServerGroup smart constructor.

Constructors

DescribeGameServerGroup'

Fields

gameServerGroupName :: Text

A unique identifier for the game server group. Use either the GameServerGroup name or ARN value.

newDescribeGameServerGroup

Arguments

:: Text
$sel:gameServerGroupName:DescribeGameServerGroup'

-> DescribeGameServerGroup

Create a value of DescribeGameServerGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:gameServerGroupName:DescribeGameServerGroup', describeGameServerGroup_gameServerGroupName - A unique identifier for the game server group. Use either the GameServerGroup name or ARN value.

Request Lenses

describeGameServerGroup_gameServerGroupName :: Lens' DescribeGameServerGroup Text

A unique identifier for the game server group. Use either the GameServerGroup name or ARN value.

Destructuring the Response

data DescribeGameServerGroupResponse

See: newDescribeGameServerGroupResponse smart constructor.

Constructors

DescribeGameServerGroupResponse'

Fields

gameServerGroup :: Maybe GameServerGroup

An object with the property settings for the requested game server group resource.

httpStatus :: Int

The response's http status code.

newDescribeGameServerGroupResponse

Arguments

:: Int
$sel:httpStatus:DescribeGameServerGroupResponse'

-> DescribeGameServerGroupResponse

Create a value of DescribeGameServerGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:gameServerGroup:DescribeGameServerGroupResponse', describeGameServerGroupResponse_gameServerGroup - An object with the property settings for the requested game server group resource.

$sel:httpStatus:DescribeGameServerGroupResponse', describeGameServerGroupResponse_httpStatus - The response's http status code.

Response Lenses

describeGameServerGroupResponse_gameServerGroup :: Lens' DescribeGameServerGroupResponse (Maybe GameServerGroup)

An object with the property settings for the requested game server group resource.

describeGameServerGroupResponse_httpStatus :: Lens' DescribeGameServerGroupResponse Int

The response's http status code.