Network.AWS.GameLift.DescribeGameServerInstances

Description

__This operation is used with the GameLift FleetIQ solution and game server groups.__

Retrieves status information about the Amazon EC2 instances associated with a GameLift FleetIQ game server group. Use this operation to detect when instances are active or not available to host new game servers. If you are looking for instance configuration information, call DescribeGameServerGroup or access the corresponding Auto Scaling group properties.

To request status for all instances in the game server group, provide a game server group ID only. To request status for specific instances, provide the game server group ID and one or more instance IDs. Use the pagination parameters to retrieve results in sequential segments. If successful, a collection of GameServerInstance objects is returned.

This operation is not designed to be called with every game server claim request; this practice can cause you to exceed your API limit, which results in errors. Instead, as a best practice, cache the results and refresh your cache no more than once every 10 seconds.

Learn more

GameLift FleetIQ Guide

Related actions

CreateGameServerGroup | ListGameServerGroups | DescribeGameServerGroup | UpdateGameServerGroup | DeleteGameServerGroup | ResumeGameServerGroup | SuspendGameServerGroup | DescribeGameServerInstances | All APIs by task

This operation returns paginated results.

Creating a Request

data DescribeGameServerInstances

See: newDescribeGameServerInstances smart constructor.

Constructors

DescribeGameServerInstances'

Fields

instanceIds :: Maybe (NonEmpty Text)

The EC2 instance IDs that you want to retrieve status on. EC2 instance IDs use a 17-character format, for example: i-1234567890abcdef0. To retrieve all instances in the game server group, leave this parameter empty.

nextToken :: Maybe Text

A token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this operation. To start at the beginning of the result set, do not specify a value.

limit :: Maybe Natural

The maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages.

gameServerGroupName :: Text

A unique identifier for the game server group. Use either the GameServerGroup name or ARN value.

newDescribeGameServerInstances

Arguments

:: Text
$sel:gameServerGroupName:DescribeGameServerInstances'

-> DescribeGameServerInstances

Create a value of DescribeGameServerInstances with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceIds:DescribeGameServerInstances', describeGameServerInstances_instanceIds - The EC2 instance IDs that you want to retrieve status on. EC2 instance IDs use a 17-character format, for example: i-1234567890abcdef0. To retrieve all instances in the game server group, leave this parameter empty.

$sel:nextToken:DescribeGameServerInstances', describeGameServerInstances_nextToken - A token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this operation. To start at the beginning of the result set, do not specify a value.

$sel:limit:DescribeGameServerInstances', describeGameServerInstances_limit - The maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages.

$sel:gameServerGroupName:DescribeGameServerInstances', describeGameServerInstances_gameServerGroupName - A unique identifier for the game server group. Use either the GameServerGroup name or ARN value.

Request Lenses

describeGameServerInstances_instanceIds :: Lens' DescribeGameServerInstances (Maybe (NonEmpty Text))

The EC2 instance IDs that you want to retrieve status on. EC2 instance IDs use a 17-character format, for example: i-1234567890abcdef0. To retrieve all instances in the game server group, leave this parameter empty.

describeGameServerInstances_nextToken :: Lens' DescribeGameServerInstances (Maybe Text)

A token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this operation. To start at the beginning of the result set, do not specify a value.

describeGameServerInstances_limit :: Lens' DescribeGameServerInstances (Maybe Natural)

The maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages.

describeGameServerInstances_gameServerGroupName :: Lens' DescribeGameServerInstances Text

A unique identifier for the game server group. Use either the GameServerGroup name or ARN value.

Destructuring the Response

data DescribeGameServerInstancesResponse

See: newDescribeGameServerInstancesResponse smart constructor.

Constructors

DescribeGameServerInstancesResponse'

Fields

nextToken :: Maybe Text

A token that indicates where to resume retrieving results on the next call to this operation. If no token is returned, these results represent the end of the list.

gameServerInstances :: Maybe [GameServerInstance]

The collection of requested game server instances.

httpStatus :: Int

The response's http status code.

newDescribeGameServerInstancesResponse

Arguments

:: Int
$sel:httpStatus:DescribeGameServerInstancesResponse'

-> DescribeGameServerInstancesResponse

Create a value of DescribeGameServerInstancesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeGameServerInstances', describeGameServerInstancesResponse_nextToken - A token that indicates where to resume retrieving results on the next call to this operation. If no token is returned, these results represent the end of the list.

$sel:gameServerInstances:DescribeGameServerInstancesResponse', describeGameServerInstancesResponse_gameServerInstances - The collection of requested game server instances.

$sel:httpStatus:DescribeGameServerInstancesResponse', describeGameServerInstancesResponse_httpStatus - The response's http status code.

Response Lenses

describeGameServerInstancesResponse_nextToken :: Lens' DescribeGameServerInstancesResponse (Maybe Text)

A token that indicates where to resume retrieving results on the next call to this operation. If no token is returned, these results represent the end of the list.

describeGameServerInstancesResponse_gameServerInstances :: Lens' DescribeGameServerInstancesResponse (Maybe [GameServerInstance])

The collection of requested game server instances.

describeGameServerInstancesResponse_httpStatus :: Lens' DescribeGameServerInstancesResponse Int

The response's http status code.