Network.AWS.GameLift.DescribeGameSessionDetails

Description

Retrieves additional game session properties, including the game session protection policy in force, a set of one or more game sessions in a specific fleet location. You can optionally filter the results by current game session status. Alternatively, use SearchGameSessions to request a set of active game sessions that are filtered by certain criteria. To retrieve all game session properties, use DescribeGameSessions.

This operation can be used in the following ways:

To retrieve details for all game sessions that are currently running on all locations in a fleet, provide a fleet or alias ID, with an optional status filter. This approach returns details from the fleet's home Region and all remote locations.

To retrieve details for all game sessions that are currently running on a specific fleet location, provide a fleet or alias ID and a location name, with optional status filter. The location can be the fleet's home Region or any remote location.

To retrieve details for a specific game session, provide the game session ID. This approach looks for the game session ID in all fleets that reside in the AWS Region defined in the request.

Use the pagination parameters to retrieve results as a set of sequential pages.

If successful, a GameSessionDetail object is returned for each game session that matches the request.

Learn more

Find a game session

Related actions

CreateGameSession | DescribeGameSessions | DescribeGameSessionDetails | SearchGameSessions | UpdateGameSession | GetGameSessionLogUrl | StartGameSessionPlacement | DescribeGameSessionPlacement | StopGameSessionPlacement | All APIs by task

This operation returns paginated results.

Creating a Request

data DescribeGameSessionDetails

Represents the input for a request operation.

See: newDescribeGameSessionDetails smart constructor.

Constructors

DescribeGameSessionDetails'

Fields

location :: Maybe Text

A fleet location to get game sessions for. You can specify a fleet's home Region or a remote location. Use the AWS Region code format, such as us-west-2.

gameSessionId :: Maybe Text

A unique identifier for the game session to retrieve.

aliasId :: Maybe Text

A unique identifier for the alias associated with the fleet to retrieve all game sessions for. You can use either the alias ID or ARN value.

nextToken :: Maybe Text

A token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this operation. To start at the beginning of the result set, do not specify a value.

statusFilter :: Maybe Text

Game session status to filter results on. Possible game session statuses include ACTIVE, TERMINATED, ACTIVATING and TERMINATING (the last two are transitory).

limit :: Maybe Natural

The maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages.

fleetId :: Maybe Text

A unique identifier for the fleet to retrieve all game sessions active on the fleet. You can use either the fleet ID or ARN value.

newDescribeGameSessionDetails :: DescribeGameSessionDetails

Create a value of DescribeGameSessionDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:location:DescribeGameSessionDetails', describeGameSessionDetails_location - A fleet location to get game sessions for. You can specify a fleet's home Region or a remote location. Use the AWS Region code format, such as us-west-2.

$sel:gameSessionId:DescribeGameSessionDetails', describeGameSessionDetails_gameSessionId - A unique identifier for the game session to retrieve.

$sel:aliasId:DescribeGameSessionDetails', describeGameSessionDetails_aliasId - A unique identifier for the alias associated with the fleet to retrieve all game sessions for. You can use either the alias ID or ARN value.

$sel:nextToken:DescribeGameSessionDetails', describeGameSessionDetails_nextToken - A token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this operation. To start at the beginning of the result set, do not specify a value.

$sel:statusFilter:DescribeGameSessionDetails', describeGameSessionDetails_statusFilter - Game session status to filter results on. Possible game session statuses include ACTIVE, TERMINATED, ACTIVATING and TERMINATING (the last two are transitory).

$sel:limit:DescribeGameSessionDetails', describeGameSessionDetails_limit - The maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages.

$sel:fleetId:DescribeGameSessionDetails', describeGameSessionDetails_fleetId - A unique identifier for the fleet to retrieve all game sessions active on the fleet. You can use either the fleet ID or ARN value.

Request Lenses

describeGameSessionDetails_location :: Lens' DescribeGameSessionDetails (Maybe Text)

A fleet location to get game sessions for. You can specify a fleet's home Region or a remote location. Use the AWS Region code format, such as us-west-2.

describeGameSessionDetails_gameSessionId :: Lens' DescribeGameSessionDetails (Maybe Text)

A unique identifier for the game session to retrieve.

describeGameSessionDetails_aliasId :: Lens' DescribeGameSessionDetails (Maybe Text)

A unique identifier for the alias associated with the fleet to retrieve all game sessions for. You can use either the alias ID or ARN value.

describeGameSessionDetails_nextToken :: Lens' DescribeGameSessionDetails (Maybe Text)

A token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this operation. To start at the beginning of the result set, do not specify a value.

describeGameSessionDetails_statusFilter :: Lens' DescribeGameSessionDetails (Maybe Text)

Game session status to filter results on. Possible game session statuses include ACTIVE, TERMINATED, ACTIVATING and TERMINATING (the last two are transitory).

describeGameSessionDetails_limit :: Lens' DescribeGameSessionDetails (Maybe Natural)

The maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages.

describeGameSessionDetails_fleetId :: Lens' DescribeGameSessionDetails (Maybe Text)

A unique identifier for the fleet to retrieve all game sessions active on the fleet. You can use either the fleet ID or ARN value.

Destructuring the Response

data DescribeGameSessionDetailsResponse

Represents the returned data in response to a request operation.

See: newDescribeGameSessionDetailsResponse smart constructor.

Constructors

DescribeGameSessionDetailsResponse'

Fields

gameSessionDetails :: Maybe [GameSessionDetail]

A collection of properties for each game session that matches the request.

nextToken :: Maybe Text

A token that indicates where to resume retrieving results on the next call to this operation. If no token is returned, these results represent the end of the list.

httpStatus :: Int

The response's http status code.

newDescribeGameSessionDetailsResponse

Arguments

:: Int
$sel:httpStatus:DescribeGameSessionDetailsResponse'

-> DescribeGameSessionDetailsResponse

Create a value of DescribeGameSessionDetailsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:gameSessionDetails:DescribeGameSessionDetailsResponse', describeGameSessionDetailsResponse_gameSessionDetails - A collection of properties for each game session that matches the request.

$sel:nextToken:DescribeGameSessionDetails', describeGameSessionDetailsResponse_nextToken - A token that indicates where to resume retrieving results on the next call to this operation. If no token is returned, these results represent the end of the list.

$sel:httpStatus:DescribeGameSessionDetailsResponse', describeGameSessionDetailsResponse_httpStatus - The response's http status code.

Response Lenses

describeGameSessionDetailsResponse_gameSessionDetails :: Lens' DescribeGameSessionDetailsResponse (Maybe [GameSessionDetail])

A collection of properties for each game session that matches the request.

describeGameSessionDetailsResponse_nextToken :: Lens' DescribeGameSessionDetailsResponse (Maybe Text)

A token that indicates where to resume retrieving results on the next call to this operation. If no token is returned, these results represent the end of the list.

describeGameSessionDetailsResponse_httpStatus :: Lens' DescribeGameSessionDetailsResponse Int

The response's http status code.