Network.AWS.GameLift.DescribeGameSessionQueues

Description

Retrieves the properties for one or more game session queues. When requesting multiple queues, use the pagination parameters to retrieve results as a set of sequential pages. If successful, a GameSessionQueue object is returned for each requested queue. When specifying a list of queues, objects are returned only for queues that currently exist in the Region.

Learn more

View Your Queues

Related actions

CreateGameSessionQueue | DescribeGameSessionQueues | UpdateGameSessionQueue | DeleteGameSessionQueue | All APIs by task

This operation returns paginated results.

Creating a Request

data DescribeGameSessionQueues

Represents the input for a request operation.

See: newDescribeGameSessionQueues smart constructor.

Constructors

DescribeGameSessionQueues'

Fields

nextToken :: Maybe Text

A token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this operation. To start at the beginning of the result set, do not specify a value.

names :: Maybe [Text]

A list of queue names to retrieve information for. You can use either the queue ID or ARN value. To request settings for all queues, leave this parameter empty.

limit :: Maybe Natural

The maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages. You can request up to 50 results.

newDescribeGameSessionQueues :: DescribeGameSessionQueues

Create a value of DescribeGameSessionQueues with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeGameSessionQueues', describeGameSessionQueues_nextToken - A token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this operation. To start at the beginning of the result set, do not specify a value.

$sel:names:DescribeGameSessionQueues', describeGameSessionQueues_names - A list of queue names to retrieve information for. You can use either the queue ID or ARN value. To request settings for all queues, leave this parameter empty.

$sel:limit:DescribeGameSessionQueues', describeGameSessionQueues_limit - The maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages. You can request up to 50 results.

Request Lenses

describeGameSessionQueues_nextToken :: Lens' DescribeGameSessionQueues (Maybe Text)

A token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this operation. To start at the beginning of the result set, do not specify a value.

describeGameSessionQueues_names :: Lens' DescribeGameSessionQueues (Maybe [Text])

A list of queue names to retrieve information for. You can use either the queue ID or ARN value. To request settings for all queues, leave this parameter empty.

describeGameSessionQueues_limit :: Lens' DescribeGameSessionQueues (Maybe Natural)

The maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages. You can request up to 50 results.

Destructuring the Response

data DescribeGameSessionQueuesResponse

Represents the returned data in response to a request operation.

See: newDescribeGameSessionQueuesResponse smart constructor.

Constructors

DescribeGameSessionQueuesResponse'

Fields

nextToken :: Maybe Text

A token that indicates where to resume retrieving results on the next call to this operation. If no token is returned, these results represent the end of the list.

gameSessionQueues :: Maybe [GameSessionQueue]

A collection of objects that describe the requested game session queues.

httpStatus :: Int

The response's http status code.

newDescribeGameSessionQueuesResponse

Arguments

:: Int
$sel:httpStatus:DescribeGameSessionQueuesResponse'

-> DescribeGameSessionQueuesResponse

Create a value of DescribeGameSessionQueuesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeGameSessionQueues', describeGameSessionQueuesResponse_nextToken - A token that indicates where to resume retrieving results on the next call to this operation. If no token is returned, these results represent the end of the list.

$sel:gameSessionQueues:DescribeGameSessionQueuesResponse', describeGameSessionQueuesResponse_gameSessionQueues - A collection of objects that describe the requested game session queues.

$sel:httpStatus:DescribeGameSessionQueuesResponse', describeGameSessionQueuesResponse_httpStatus - The response's http status code.

Response Lenses

describeGameSessionQueuesResponse_nextToken :: Lens' DescribeGameSessionQueuesResponse (Maybe Text)

A token that indicates where to resume retrieving results on the next call to this operation. If no token is returned, these results represent the end of the list.

describeGameSessionQueuesResponse_gameSessionQueues :: Lens' DescribeGameSessionQueuesResponse (Maybe [GameSessionQueue])

A collection of objects that describe the requested game session queues.

describeGameSessionQueuesResponse_httpStatus :: Lens' DescribeGameSessionQueuesResponse Int

The response's http status code.