Network.AWS.GameLift.DescribeGameSessions

Description

Retrieves a set of one or more game sessions in a specific fleet location. You can optionally filter the results by current game session status. Alternatively, use SearchGameSessions to request a set of active game sessions that are filtered by certain criteria. To retrieve the protection policy for game sessions, use DescribeGameSessionDetails.

This operation can be used in the following ways:

To retrieve all game sessions that are currently running on all locations in a fleet, provide a fleet or alias ID, with an optional status filter. This approach returns all game sessions in the fleet's home Region and all remote locations.

To retrieve all game sessions that are currently running on a specific fleet location, provide a fleet or alias ID and a location name, with optional status filter. The location can be the fleet's home Region or any remote location.

To retrieve a specific game session, provide the game session ID. This approach looks for the game session ID in all fleets that reside in the AWS Region defined in the request.

Use the pagination parameters to retrieve results as a set of sequential pages.

If successful, a GameSession object is returned for each game session that matches the request.

Available in GameLift Local.

Learn more

Find a game session

Related actions

CreateGameSession | DescribeGameSessions | DescribeGameSessionDetails | SearchGameSessions | UpdateGameSession | GetGameSessionLogUrl | StartGameSessionPlacement | DescribeGameSessionPlacement | StopGameSessionPlacement | All APIs by task

This operation returns paginated results.

Creating a Request

data DescribeGameSessions

Represents the input for a request operation.

See: newDescribeGameSessions smart constructor.

Constructors

DescribeGameSessions'

Fields

nextToken :: Maybe Text

A token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this operation. To start at the beginning of the result set, do not specify a value.

fleetId :: Maybe Text

A unique identifier for the fleet to retrieve game sessions for. You can use either the fleet ID or ARN value.

gameSessionId :: Maybe Text

A unique identifier for the game session to retrieve.

statusFilter :: Maybe Text

Game session status to filter results on. You can filter on the following states: ACTIVE, TERMINATED, ACTIVATING, and TERMINATING. The last two are transitory and used for only very brief periods of time.

aliasId :: Maybe Text

A unique identifier for the alias associated with the fleet to retrieve game sessions for. You can use either the alias ID or ARN value.

location :: Maybe Text

A fleet location to get game session details for. You can specify a fleet's home Region or a remote location. Use the AWS Region code format, such as us-west-2.

limit :: Maybe Natural

The maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages.

newDescribeGameSessions :: DescribeGameSessions

Create a value of DescribeGameSessions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeGameSessions', describeGameSessions_nextToken - A token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this operation. To start at the beginning of the result set, do not specify a value.

$sel:fleetId:DescribeGameSessions', describeGameSessions_fleetId - A unique identifier for the fleet to retrieve game sessions for. You can use either the fleet ID or ARN value.

$sel:gameSessionId:DescribeGameSessions', describeGameSessions_gameSessionId - A unique identifier for the game session to retrieve.

$sel:statusFilter:DescribeGameSessions', describeGameSessions_statusFilter - Game session status to filter results on. You can filter on the following states: ACTIVE, TERMINATED, ACTIVATING, and TERMINATING. The last two are transitory and used for only very brief periods of time.

$sel:aliasId:DescribeGameSessions', describeGameSessions_aliasId - A unique identifier for the alias associated with the fleet to retrieve game sessions for. You can use either the alias ID or ARN value.

$sel:location:DescribeGameSessions', describeGameSessions_location - A fleet location to get game session details for. You can specify a fleet's home Region or a remote location. Use the AWS Region code format, such as us-west-2.

$sel:limit:DescribeGameSessions', describeGameSessions_limit - The maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages.

Request Lenses

describeGameSessions_nextToken :: Lens' DescribeGameSessions (Maybe Text)

A token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this operation. To start at the beginning of the result set, do not specify a value.

describeGameSessions_fleetId :: Lens' DescribeGameSessions (Maybe Text)

A unique identifier for the fleet to retrieve game sessions for. You can use either the fleet ID or ARN value.

describeGameSessions_gameSessionId :: Lens' DescribeGameSessions (Maybe Text)

A unique identifier for the game session to retrieve.

describeGameSessions_statusFilter :: Lens' DescribeGameSessions (Maybe Text)

Game session status to filter results on. You can filter on the following states: ACTIVE, TERMINATED, ACTIVATING, and TERMINATING. The last two are transitory and used for only very brief periods of time.

describeGameSessions_aliasId :: Lens' DescribeGameSessions (Maybe Text)

A unique identifier for the alias associated with the fleet to retrieve game sessions for. You can use either the alias ID or ARN value.

describeGameSessions_location :: Lens' DescribeGameSessions (Maybe Text)

A fleet location to get game session details for. You can specify a fleet's home Region or a remote location. Use the AWS Region code format, such as us-west-2.

describeGameSessions_limit :: Lens' DescribeGameSessions (Maybe Natural)

The maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages.

Destructuring the Response

data DescribeGameSessionsResponse

Represents the returned data in response to a request operation.

See: newDescribeGameSessionsResponse smart constructor.

Constructors

DescribeGameSessionsResponse'

Fields

nextToken :: Maybe Text

A token that indicates where to resume retrieving results on the next call to this operation. If no token is returned, these results represent the end of the list.

gameSessions :: Maybe [GameSession]

A collection of properties for each game session that matches the request.

httpStatus :: Int

The response's http status code.

newDescribeGameSessionsResponse

Arguments

:: Int
$sel:httpStatus:DescribeGameSessionsResponse'

-> DescribeGameSessionsResponse

Create a value of DescribeGameSessionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeGameSessions', describeGameSessionsResponse_nextToken - A token that indicates where to resume retrieving results on the next call to this operation. If no token is returned, these results represent the end of the list.

$sel:gameSessions:DescribeGameSessionsResponse', describeGameSessionsResponse_gameSessions - A collection of properties for each game session that matches the request.

$sel:httpStatus:DescribeGameSessionsResponse', describeGameSessionsResponse_httpStatus - The response's http status code.

Response Lenses

describeGameSessionsResponse_nextToken :: Lens' DescribeGameSessionsResponse (Maybe Text)

A token that indicates where to resume retrieving results on the next call to this operation. If no token is returned, these results represent the end of the list.

describeGameSessionsResponse_gameSessions :: Lens' DescribeGameSessionsResponse (Maybe [GameSession])

A collection of properties for each game session that matches the request.

describeGameSessionsResponse_httpStatus :: Lens' DescribeGameSessionsResponse Int

The response's http status code.