Network.AWS.GameLift.DescribeMatchmakingRuleSets

Description

Retrieves the details for FlexMatch matchmaking rule sets. You can request all existing rule sets for the Region, or provide a list of one or more rule set names. When requesting multiple items, use the pagination parameters to retrieve results as a set of sequential pages. If successful, a rule set is returned for each requested name.

Learn more

Build a rule set

Related actions

CreateMatchmakingConfiguration | DescribeMatchmakingConfigurations | UpdateMatchmakingConfiguration | DeleteMatchmakingConfiguration | CreateMatchmakingRuleSet | DescribeMatchmakingRuleSets | ValidateMatchmakingRuleSet | DeleteMatchmakingRuleSet | All APIs by task

This operation returns paginated results.

Creating a Request

data DescribeMatchmakingRuleSets

Represents the input for a request operation.

See: newDescribeMatchmakingRuleSets smart constructor.

Constructors

DescribeMatchmakingRuleSets'

Fields

nextToken :: Maybe Text

A token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this operation. To start at the beginning of the result set, do not specify a value.

names :: Maybe (NonEmpty Text)

A list of one or more matchmaking rule set names to retrieve details for. (Note: The rule set name is different from the optional "name" field in the rule set body.) You can use either the rule set name or ARN value.

limit :: Maybe Natural

The maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages.

newDescribeMatchmakingRuleSets :: DescribeMatchmakingRuleSets

Create a value of DescribeMatchmakingRuleSets with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeMatchmakingRuleSets', describeMatchmakingRuleSets_nextToken - A token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this operation. To start at the beginning of the result set, do not specify a value.

$sel:names:DescribeMatchmakingRuleSets', describeMatchmakingRuleSets_names - A list of one or more matchmaking rule set names to retrieve details for. (Note: The rule set name is different from the optional "name" field in the rule set body.) You can use either the rule set name or ARN value.

$sel:limit:DescribeMatchmakingRuleSets', describeMatchmakingRuleSets_limit - The maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages.

Request Lenses

describeMatchmakingRuleSets_nextToken :: Lens' DescribeMatchmakingRuleSets (Maybe Text)

A token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this operation. To start at the beginning of the result set, do not specify a value.

describeMatchmakingRuleSets_names :: Lens' DescribeMatchmakingRuleSets (Maybe (NonEmpty Text))

A list of one or more matchmaking rule set names to retrieve details for. (Note: The rule set name is different from the optional "name" field in the rule set body.) You can use either the rule set name or ARN value.

describeMatchmakingRuleSets_limit :: Lens' DescribeMatchmakingRuleSets (Maybe Natural)

The maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages.

Destructuring the Response

data DescribeMatchmakingRuleSetsResponse

Represents the returned data in response to a request operation.

See: newDescribeMatchmakingRuleSetsResponse smart constructor.

Constructors

DescribeMatchmakingRuleSetsResponse'

Fields

nextToken :: Maybe Text

A token that indicates where to resume retrieving results on the next call to this operation. If no token is returned, these results represent the end of the list.

httpStatus :: Int

The response's http status code.

ruleSets :: [MatchmakingRuleSet]

A collection of requested matchmaking rule set objects.

newDescribeMatchmakingRuleSetsResponse

Arguments

:: Int
$sel:httpStatus:DescribeMatchmakingRuleSetsResponse'

-> DescribeMatchmakingRuleSetsResponse

Create a value of DescribeMatchmakingRuleSetsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeMatchmakingRuleSets', describeMatchmakingRuleSetsResponse_nextToken - A token that indicates where to resume retrieving results on the next call to this operation. If no token is returned, these results represent the end of the list.

$sel:httpStatus:DescribeMatchmakingRuleSetsResponse', describeMatchmakingRuleSetsResponse_httpStatus - The response's http status code.

$sel:ruleSets:DescribeMatchmakingRuleSetsResponse', describeMatchmakingRuleSetsResponse_ruleSets - A collection of requested matchmaking rule set objects.

Response Lenses

describeMatchmakingRuleSetsResponse_nextToken :: Lens' DescribeMatchmakingRuleSetsResponse (Maybe Text)

A token that indicates where to resume retrieving results on the next call to this operation. If no token is returned, these results represent the end of the list.

describeMatchmakingRuleSetsResponse_httpStatus :: Lens' DescribeMatchmakingRuleSetsResponse Int

The response's http status code.

describeMatchmakingRuleSetsResponse_ruleSets :: Lens' DescribeMatchmakingRuleSetsResponse [MatchmakingRuleSet]

A collection of requested matchmaking rule set objects.