Network.AWS.GameLift.DescribePlayerSessions

Description

Retrieves properties for one or more player sessions.

This action can be used in the following ways:

To retrieve a specific player session, provide the player session ID only.

To retrieve all player sessions in a game session, provide the game session ID only.

To retrieve all player sessions for a specific player, provide a player ID only.

To request player sessions, specify either a player session ID, game session ID, or player ID. You can filter this request by player session status. Use the pagination parameters to retrieve results as a set of sequential pages.

If successful, a PlayerSession object is returned for each session that matches the request.

Available in Amazon GameLift Local.

Related actions

CreatePlayerSession | CreatePlayerSessions | DescribePlayerSessions | StartGameSessionPlacement | DescribeGameSessionPlacement | All APIs by task

This operation returns paginated results.

Creating a Request

data DescribePlayerSessions

Represents the input for a request operation.

See: newDescribePlayerSessions smart constructor.

Constructors

DescribePlayerSessions'

Fields

gameSessionId :: Maybe Text

A unique identifier for the game session to retrieve player sessions for.

nextToken :: Maybe Text

A token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this operation. To start at the beginning of the result set, do not specify a value. If a player session ID is specified, this parameter is ignored.

limit :: Maybe Natural

The maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages. If a player session ID is specified, this parameter is ignored.

playerSessionId :: Maybe Text

A unique identifier for a player session to retrieve.

playerId :: Maybe Text

A unique identifier for a player to retrieve player sessions for.

playerSessionStatusFilter :: Maybe Text

Player session status to filter results on.

Possible player session statuses include the following:

RESERVED -- The player session request has been received, but the player has not yet connected to the server process and/or been validated.

ACTIVE -- The player has been validated by the server process and is currently connected.

COMPLETED -- The player connection has been dropped.

TIMEDOUT -- A player session request was received, but the player did not connect and/or was not validated within the timeout limit (60 seconds).

newDescribePlayerSessions :: DescribePlayerSessions

Create a value of DescribePlayerSessions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:gameSessionId:DescribePlayerSessions', describePlayerSessions_gameSessionId - A unique identifier for the game session to retrieve player sessions for.

$sel:nextToken:DescribePlayerSessions', describePlayerSessions_nextToken - A token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this operation. To start at the beginning of the result set, do not specify a value. If a player session ID is specified, this parameter is ignored.

$sel:limit:DescribePlayerSessions', describePlayerSessions_limit - The maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages. If a player session ID is specified, this parameter is ignored.

$sel:playerSessionId:DescribePlayerSessions', describePlayerSessions_playerSessionId - A unique identifier for a player session to retrieve.

$sel:playerId:DescribePlayerSessions', describePlayerSessions_playerId - A unique identifier for a player to retrieve player sessions for.

$sel:playerSessionStatusFilter:DescribePlayerSessions', describePlayerSessions_playerSessionStatusFilter - Player session status to filter results on.

Possible player session statuses include the following:

RESERVED -- The player session request has been received, but the player has not yet connected to the server process and/or been validated.

ACTIVE -- The player has been validated by the server process and is currently connected.

COMPLETED -- The player connection has been dropped.

TIMEDOUT -- A player session request was received, but the player did not connect and/or was not validated within the timeout limit (60 seconds).

Request Lenses

describePlayerSessions_gameSessionId :: Lens' DescribePlayerSessions (Maybe Text)

A unique identifier for the game session to retrieve player sessions for.

describePlayerSessions_nextToken :: Lens' DescribePlayerSessions (Maybe Text)

A token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this operation. To start at the beginning of the result set, do not specify a value. If a player session ID is specified, this parameter is ignored.

describePlayerSessions_limit :: Lens' DescribePlayerSessions (Maybe Natural)

The maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages. If a player session ID is specified, this parameter is ignored.

describePlayerSessions_playerSessionId :: Lens' DescribePlayerSessions (Maybe Text)

A unique identifier for a player session to retrieve.

describePlayerSessions_playerId :: Lens' DescribePlayerSessions (Maybe Text)

A unique identifier for a player to retrieve player sessions for.

describePlayerSessions_playerSessionStatusFilter :: Lens' DescribePlayerSessions (Maybe Text)

Player session status to filter results on.

Possible player session statuses include the following:

RESERVED -- The player session request has been received, but the player has not yet connected to the server process and/or been validated.

ACTIVE -- The player has been validated by the server process and is currently connected.

COMPLETED -- The player connection has been dropped.

TIMEDOUT -- A player session request was received, but the player did not connect and/or was not validated within the timeout limit (60 seconds).

Destructuring the Response

data DescribePlayerSessionsResponse

Represents the returned data in response to a request operation.

See: newDescribePlayerSessionsResponse smart constructor.

Constructors

DescribePlayerSessionsResponse'

Fields

nextToken :: Maybe Text

A token that indicates where to resume retrieving results on the next call to this operation. If no token is returned, these results represent the end of the list.

playerSessions :: Maybe [PlayerSession]

A collection of objects containing properties for each player session that matches the request.

httpStatus :: Int

The response's http status code.

newDescribePlayerSessionsResponse

Arguments

:: Int
$sel:httpStatus:DescribePlayerSessionsResponse'

-> DescribePlayerSessionsResponse

Create a value of DescribePlayerSessionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribePlayerSessions', describePlayerSessionsResponse_nextToken - A token that indicates where to resume retrieving results on the next call to this operation. If no token is returned, these results represent the end of the list.

$sel:playerSessions:DescribePlayerSessionsResponse', describePlayerSessionsResponse_playerSessions - A collection of objects containing properties for each player session that matches the request.

$sel:httpStatus:DescribePlayerSessionsResponse', describePlayerSessionsResponse_httpStatus - The response's http status code.

Response Lenses

describePlayerSessionsResponse_nextToken :: Lens' DescribePlayerSessionsResponse (Maybe Text)

A token that indicates where to resume retrieving results on the next call to this operation. If no token is returned, these results represent the end of the list.

describePlayerSessionsResponse_playerSessions :: Lens' DescribePlayerSessionsResponse (Maybe [PlayerSession])

A collection of objects containing properties for each player session that matches the request.

describePlayerSessionsResponse_httpStatus :: Lens' DescribePlayerSessionsResponse Int

The response's http status code.