Network.AWS.GameLift.DescribeScalingPolicies

Description

Retrieves all scaling policies applied to a fleet.

To get a fleet's scaling policies, specify the fleet ID. You can filter this request by policy status, such as to retrieve only active scaling policies. Use the pagination parameters to retrieve results as a set of sequential pages. If successful, set of ScalingPolicy objects is returned for the fleet.

A fleet may have all of its scaling policies suspended (StopFleetActions). This operation does not affect the status of the scaling policies, which remains ACTIVE. To see whether a fleet's scaling policies are in force or suspended, call DescribeFleetAttributes and check the stopped actions.

Related actions

DescribeFleetCapacity | UpdateFleetCapacity | DescribeEC2InstanceLimits | PutScalingPolicy | DescribeScalingPolicies | DeleteScalingPolicy | StopFleetActions | StartFleetActions | All APIs by task

This operation returns paginated results.

Creating a Request

data DescribeScalingPolicies

Represents the input for a request operation.

See: newDescribeScalingPolicies smart constructor.

Constructors

DescribeScalingPolicies'

Fields

location :: Maybe Text

CONTENT TODO

nextToken :: Maybe Text

A token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this operation. To start at the beginning of the result set, do not specify a value.

statusFilter :: Maybe ScalingStatusType

Scaling policy status to filter results on. A scaling policy is only in force when in an ACTIVE status.

ACTIVE -- The scaling policy is currently in force.

UPDATEREQUESTED -- A request to update the scaling policy has been received.

UPDATING -- A change is being made to the scaling policy.

DELETEREQUESTED -- A request to delete the scaling policy has been received.

DELETING -- The scaling policy is being deleted.

DELETED -- The scaling policy has been deleted.

ERROR -- An error occurred in creating the policy. It should be removed and recreated.

limit :: Maybe Natural

The maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages.

fleetId :: Text

A unique identifier for the fleet to retrieve scaling policies for. You can use either the fleet ID or ARN value.

newDescribeScalingPolicies

Arguments

:: Text
$sel:fleetId:DescribeScalingPolicies'

-> DescribeScalingPolicies

Create a value of DescribeScalingPolicies with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:location:DescribeScalingPolicies', describeScalingPolicies_location - CONTENT TODO

$sel:nextToken:DescribeScalingPolicies', describeScalingPolicies_nextToken - A token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this operation. To start at the beginning of the result set, do not specify a value.

$sel:statusFilter:DescribeScalingPolicies', describeScalingPolicies_statusFilter - Scaling policy status to filter results on. A scaling policy is only in force when in an ACTIVE status.

ACTIVE -- The scaling policy is currently in force.

UPDATEREQUESTED -- A request to update the scaling policy has been received.

UPDATING -- A change is being made to the scaling policy.

DELETEREQUESTED -- A request to delete the scaling policy has been received.

DELETING -- The scaling policy is being deleted.

DELETED -- The scaling policy has been deleted.

ERROR -- An error occurred in creating the policy. It should be removed and recreated.

$sel:limit:DescribeScalingPolicies', describeScalingPolicies_limit - The maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages.

$sel:fleetId:DescribeScalingPolicies', describeScalingPolicies_fleetId - A unique identifier for the fleet to retrieve scaling policies for. You can use either the fleet ID or ARN value.

Request Lenses

describeScalingPolicies_location :: Lens' DescribeScalingPolicies (Maybe Text)

CONTENT TODO

describeScalingPolicies_nextToken :: Lens' DescribeScalingPolicies (Maybe Text)

A token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this operation. To start at the beginning of the result set, do not specify a value.

describeScalingPolicies_statusFilter :: Lens' DescribeScalingPolicies (Maybe ScalingStatusType)

Scaling policy status to filter results on. A scaling policy is only in force when in an ACTIVE status.

ACTIVE -- The scaling policy is currently in force.

UPDATEREQUESTED -- A request to update the scaling policy has been received.

UPDATING -- A change is being made to the scaling policy.

DELETEREQUESTED -- A request to delete the scaling policy has been received.

DELETING -- The scaling policy is being deleted.

DELETED -- The scaling policy has been deleted.

ERROR -- An error occurred in creating the policy. It should be removed and recreated.

describeScalingPolicies_limit :: Lens' DescribeScalingPolicies (Maybe Natural)

The maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages.

describeScalingPolicies_fleetId :: Lens' DescribeScalingPolicies Text

A unique identifier for the fleet to retrieve scaling policies for. You can use either the fleet ID or ARN value.

Destructuring the Response

data DescribeScalingPoliciesResponse

Represents the returned data in response to a request operation.

See: newDescribeScalingPoliciesResponse smart constructor.

Constructors

DescribeScalingPoliciesResponse'

Fields

nextToken :: Maybe Text

A token that indicates where to resume retrieving results on the next call to this operation. If no token is returned, these results represent the end of the list.

scalingPolicies :: Maybe [ScalingPolicy]

A collection of objects containing the scaling policies matching the request.

httpStatus :: Int

The response's http status code.

newDescribeScalingPoliciesResponse

Arguments

:: Int
$sel:httpStatus:DescribeScalingPoliciesResponse'

-> DescribeScalingPoliciesResponse

Create a value of DescribeScalingPoliciesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeScalingPolicies', describeScalingPoliciesResponse_nextToken - A token that indicates where to resume retrieving results on the next call to this operation. If no token is returned, these results represent the end of the list.

$sel:scalingPolicies:DescribeScalingPoliciesResponse', describeScalingPoliciesResponse_scalingPolicies - A collection of objects containing the scaling policies matching the request.

$sel:httpStatus:DescribeScalingPoliciesResponse', describeScalingPoliciesResponse_httpStatus - The response's http status code.

Response Lenses

describeScalingPoliciesResponse_nextToken :: Lens' DescribeScalingPoliciesResponse (Maybe Text)

A token that indicates where to resume retrieving results on the next call to this operation. If no token is returned, these results represent the end of the list.

describeScalingPoliciesResponse_scalingPolicies :: Lens' DescribeScalingPoliciesResponse (Maybe [ScalingPolicy])

A collection of objects containing the scaling policies matching the request.

describeScalingPoliciesResponse_httpStatus :: Lens' DescribeScalingPoliciesResponse Int

The response's http status code.