Network.AWS.GameLift.GetInstanceAccess

Description

Requests remote access to a fleet instance. Remote access is useful for debugging, gathering benchmarking data, or observing activity in real time.

To remotely access an instance, you need credentials that match the operating system of the instance. For a Windows instance, GameLift returns a user name and password as strings for use with a Windows Remote Desktop client. For a Linux instance, GameLift returns a user name and RSA private key, also as strings, for use with an SSH client. The private key must be saved in the proper format to a .pem file before using. If you're making this request using the AWS CLI, saving the secret can be handled as part of the GetInstanceAccess request, as shown in one of the examples for this operation.

To request access to a specific instance, specify the IDs of both the instance and the fleet it belongs to. You can retrieve a fleet's instance IDs by calling DescribeInstances. If successful, an InstanceAccess object is returned that contains the instance's IP address and a set of credentials.

Learn more

Remotely Access Fleet Instances

Debug Fleet Issues

Related actions

DescribeInstances | GetInstanceAccess | DescribeEC2InstanceLimits | All APIs by task

Creating a Request

data GetInstanceAccess

Represents the input for a request operation.

See: newGetInstanceAccess smart constructor.

Constructors

GetInstanceAccess'

Fields

fleetId :: Text

A unique identifier for the fleet that contains the instance you want access to. You can use either the fleet ID or ARN value. The fleet can be in any of the following statuses: ACTIVATING, ACTIVE, or ERROR. Fleets with an ERROR status may be accessible for a short time before they are deleted.

instanceId :: Text

A unique identifier for the instance you want to get access to. You can access an instance in any status.

newGetInstanceAccess

Arguments

:: Text
$sel:fleetId:GetInstanceAccess'

-> Text
$sel:instanceId:GetInstanceAccess'

-> GetInstanceAccess

Create a value of GetInstanceAccess with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fleetId:GetInstanceAccess', getInstanceAccess_fleetId - A unique identifier for the fleet that contains the instance you want access to. You can use either the fleet ID or ARN value. The fleet can be in any of the following statuses: ACTIVATING, ACTIVE, or ERROR. Fleets with an ERROR status may be accessible for a short time before they are deleted.

$sel:instanceId:GetInstanceAccess', getInstanceAccess_instanceId - A unique identifier for the instance you want to get access to. You can access an instance in any status.

Request Lenses

getInstanceAccess_fleetId :: Lens' GetInstanceAccess Text

A unique identifier for the fleet that contains the instance you want access to. You can use either the fleet ID or ARN value. The fleet can be in any of the following statuses: ACTIVATING, ACTIVE, or ERROR. Fleets with an ERROR status may be accessible for a short time before they are deleted.

getInstanceAccess_instanceId :: Lens' GetInstanceAccess Text

A unique identifier for the instance you want to get access to. You can access an instance in any status.

Destructuring the Response

data GetInstanceAccessResponse

Represents the returned data in response to a request operation.

See: newGetInstanceAccessResponse smart constructor.

Constructors

GetInstanceAccessResponse'

Fields

instanceAccess :: Maybe InstanceAccess

The connection information for a fleet instance, including IP address and access credentials.

httpStatus :: Int

The response's http status code.

newGetInstanceAccessResponse

Arguments

:: Int
$sel:httpStatus:GetInstanceAccessResponse'

-> GetInstanceAccessResponse

Create a value of GetInstanceAccessResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceAccess:GetInstanceAccessResponse', getInstanceAccessResponse_instanceAccess - The connection information for a fleet instance, including IP address and access credentials.

$sel:httpStatus:GetInstanceAccessResponse', getInstanceAccessResponse_httpStatus - The response's http status code.

Response Lenses

getInstanceAccessResponse_instanceAccess :: Lens' GetInstanceAccessResponse (Maybe InstanceAccess)

The connection information for a fleet instance, including IP address and access credentials.

getInstanceAccessResponse_httpStatus :: Lens' GetInstanceAccessResponse Int

The response's http status code.