Network.AWS.GameLift.ListScripts

Description

Retrieves script records for all Realtime scripts that are associated with the AWS account in use.

Learn more

Amazon GameLift Realtime Servers

Related actions

CreateScript | ListScripts | DescribeScript | UpdateScript | DeleteScript | All APIs by task

This operation returns paginated results.

Creating a Request

data ListScripts

See: newListScripts smart constructor.

Constructors

ListScripts'

Fields

nextToken :: Maybe Text

A token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this operation. To start at the beginning of the result set, do not specify a value.

limit :: Maybe Natural

The maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages.

newListScripts :: ListScripts

Create a value of ListScripts with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListScripts', listScripts_nextToken - A token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this operation. To start at the beginning of the result set, do not specify a value.

$sel:limit:ListScripts', listScripts_limit - The maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages.

Request Lenses

listScripts_nextToken :: Lens' ListScripts (Maybe Text)

A token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this operation. To start at the beginning of the result set, do not specify a value.

listScripts_limit :: Lens' ListScripts (Maybe Natural)

The maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages.

Destructuring the Response

data ListScriptsResponse

See: newListScriptsResponse smart constructor.

Constructors

ListScriptsResponse'

Fields

scripts :: Maybe [Script]

A set of properties describing the requested script.

nextToken :: Maybe Text

A token that indicates where to resume retrieving results on the next call to this operation. If no token is returned, these results represent the end of the list.

httpStatus :: Int

The response's http status code.

newListScriptsResponse

Arguments

:: Int
$sel:httpStatus:ListScriptsResponse'

-> ListScriptsResponse

Create a value of ListScriptsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:scripts:ListScriptsResponse', listScriptsResponse_scripts - A set of properties describing the requested script.

$sel:nextToken:ListScripts', listScriptsResponse_nextToken - A token that indicates where to resume retrieving results on the next call to this operation. If no token is returned, these results represent the end of the list.

$sel:httpStatus:ListScriptsResponse', listScriptsResponse_httpStatus - The response's http status code.

Response Lenses

listScriptsResponse_scripts :: Lens' ListScriptsResponse (Maybe [Script])

A set of properties describing the requested script.

listScriptsResponse_nextToken :: Lens' ListScriptsResponse (Maybe Text)

A token that indicates where to resume retrieving results on the next call to this operation. If no token is returned, these results represent the end of the list.

listScriptsResponse_httpStatus :: Lens' ListScriptsResponse Int

The response's http status code.