Network.AWS.GameLift.RegisterGameServer

Description

__This operation is used with the GameLift FleetIQ solution and game server groups.__

Creates a new game server resource and notifies GameLift FleetIQ that the game server is ready to host gameplay and players. This operation is called by a game server process that is running on an instance in a game server group. Registering game servers enables GameLift FleetIQ to track available game servers and enables game clients and services to claim a game server for a new game session.

To register a game server, identify the game server group and instance where the game server is running, and provide a unique identifier for the game server. You can also include connection and game server data. When a game client or service requests a game server by calling ClaimGameServer, this information is returned in the response.

Once a game server is successfully registered, it is put in status AVAILABLE. A request to register a game server may fail if the instance it is running on is in the process of shutting down as part of instance balancing or scale-down activity.

Learn more

GameLift FleetIQ Guide

Related actions

RegisterGameServer | ListGameServers | ClaimGameServer | DescribeGameServer | UpdateGameServer | DeregisterGameServer | All APIs by task

Creating a Request

data RegisterGameServer

See: newRegisterGameServer smart constructor.

Constructors

RegisterGameServer'

Fields

gameServerData :: Maybe Text

A set of custom game server properties, formatted as a single string value. This data is passed to a game client or service when it requests information on game servers using ListGameServers or ClaimGameServer.

connectionInfo :: Maybe Text

Information that is needed to make inbound client connections to the game server. This might include the IP address and port, DNS name, and other information.

gameServerGroupName :: Text

A unique identifier for the game server group where the game server is running. Use either the GameServerGroup name or ARN value.

gameServerId :: Text

A custom string that uniquely identifies the game server to register. Game server IDs are developer-defined and must be unique across all game server groups in your AWS account.

instanceId :: Text

The unique identifier for the instance where the game server is running. This ID is available in the instance metadata. EC2 instance IDs use a 17-character format, for example: i-1234567890abcdef0.

newRegisterGameServer

Arguments

:: Text
$sel:gameServerGroupName:RegisterGameServer'

-> Text
$sel:gameServerId:RegisterGameServer'

-> Text
$sel:instanceId:RegisterGameServer'

-> RegisterGameServer

Create a value of RegisterGameServer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:gameServerData:RegisterGameServer', registerGameServer_gameServerData - A set of custom game server properties, formatted as a single string value. This data is passed to a game client or service when it requests information on game servers using ListGameServers or ClaimGameServer.

$sel:connectionInfo:RegisterGameServer', registerGameServer_connectionInfo - Information that is needed to make inbound client connections to the game server. This might include the IP address and port, DNS name, and other information.

$sel:gameServerGroupName:RegisterGameServer', registerGameServer_gameServerGroupName - A unique identifier for the game server group where the game server is running. Use either the GameServerGroup name or ARN value.

$sel:gameServerId:RegisterGameServer', registerGameServer_gameServerId - A custom string that uniquely identifies the game server to register. Game server IDs are developer-defined and must be unique across all game server groups in your AWS account.

$sel:instanceId:RegisterGameServer', registerGameServer_instanceId - The unique identifier for the instance where the game server is running. This ID is available in the instance metadata. EC2 instance IDs use a 17-character format, for example: i-1234567890abcdef0.

Request Lenses

registerGameServer_gameServerData :: Lens' RegisterGameServer (Maybe Text)

A set of custom game server properties, formatted as a single string value. This data is passed to a game client or service when it requests information on game servers using ListGameServers or ClaimGameServer.

registerGameServer_connectionInfo :: Lens' RegisterGameServer (Maybe Text)

Information that is needed to make inbound client connections to the game server. This might include the IP address and port, DNS name, and other information.

registerGameServer_gameServerGroupName :: Lens' RegisterGameServer Text

A unique identifier for the game server group where the game server is running. Use either the GameServerGroup name or ARN value.

registerGameServer_gameServerId :: Lens' RegisterGameServer Text

A custom string that uniquely identifies the game server to register. Game server IDs are developer-defined and must be unique across all game server groups in your AWS account.

registerGameServer_instanceId :: Lens' RegisterGameServer Text

The unique identifier for the instance where the game server is running. This ID is available in the instance metadata. EC2 instance IDs use a 17-character format, for example: i-1234567890abcdef0.

Destructuring the Response

data RegisterGameServerResponse

See: newRegisterGameServerResponse smart constructor.

Constructors

RegisterGameServerResponse'

Fields

gameServer :: Maybe GameServer

Object that describes the newly registered game server.

httpStatus :: Int

The response's http status code.

newRegisterGameServerResponse

Arguments

:: Int
$sel:httpStatus:RegisterGameServerResponse'

-> RegisterGameServerResponse

Create a value of RegisterGameServerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:gameServer:RegisterGameServerResponse', registerGameServerResponse_gameServer - Object that describes the newly registered game server.

$sel:httpStatus:RegisterGameServerResponse', registerGameServerResponse_httpStatus - The response's http status code.

Response Lenses

registerGameServerResponse_gameServer :: Lens' RegisterGameServerResponse (Maybe GameServer)

Object that describes the newly registered game server.

registerGameServerResponse_httpStatus :: Lens' RegisterGameServerResponse Int

The response's http status code.