Network.AWS.GameLift.StartFleetActions

Description

Resumes certain types of activity on fleet instances that were suspended with StopFleetActions. For multi-location fleets, fleet actions are managed separately for each location. Currently, this operation is used to restart a fleet's auto-scaling activity.

This operation can be used in the following ways:

To restart actions on instances in the fleet's home Region, provide a fleet ID and the type of actions to resume.

To restart actions on instances in one of the fleet's remote locations, provide a fleet ID, a location name, and the type of actions to resume.

If successful, GameLift once again initiates scaling events as triggered by the fleet's scaling policies. If actions on the fleet location were never stopped, this operation will have no effect. You can view a fleet's stopped actions using DescribeFleetAttributes or DescribeFleetLocationAttributes.

Learn more

Setting up GameLift fleets

Related actions

CreateFleet | UpdateFleetCapacity | PutScalingPolicy | DescribeEC2InstanceLimits | DescribeFleetAttributes | DescribeFleetLocationAttributes | UpdateFleetAttributes | StopFleetActions | DeleteFleet | All APIs by task

Creating a Request

data StartFleetActions

Represents the input for a request operation.

See: newStartFleetActions smart constructor.

Constructors

StartFleetActions'

Fields

location :: Maybe Text

The fleet location to restart fleet actions for. Specify a location in the form of an AWS Region code, such as us-west-2.

fleetId :: Text

A unique identifier for the fleet to restart actions on. You can use either the fleet ID or ARN value.

actions :: NonEmpty FleetAction

List of actions to restart on the fleet.

newStartFleetActions

Arguments

:: Text
$sel:fleetId:StartFleetActions'

-> NonEmpty FleetAction
$sel:actions:StartFleetActions'

-> StartFleetActions

Create a value of StartFleetActions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:location:StartFleetActions', startFleetActions_location - The fleet location to restart fleet actions for. Specify a location in the form of an AWS Region code, such as us-west-2.

$sel:fleetId:StartFleetActions', startFleetActions_fleetId - A unique identifier for the fleet to restart actions on. You can use either the fleet ID or ARN value.

$sel:actions:StartFleetActions', startFleetActions_actions - List of actions to restart on the fleet.

Request Lenses

startFleetActions_location :: Lens' StartFleetActions (Maybe Text)

The fleet location to restart fleet actions for. Specify a location in the form of an AWS Region code, such as us-west-2.

startFleetActions_fleetId :: Lens' StartFleetActions Text

A unique identifier for the fleet to restart actions on. You can use either the fleet ID or ARN value.

startFleetActions_actions :: Lens' StartFleetActions (NonEmpty FleetAction)

List of actions to restart on the fleet.

Destructuring the Response

data StartFleetActionsResponse

Represents the returned data in response to a request operation.

See: newStartFleetActionsResponse smart constructor.

Constructors

StartFleetActionsResponse'

Fields

fleetId :: Maybe Text

A unique identifier for the fleet to restart actions on.

fleetArn :: Maybe Text

The Amazon Resource Name (ARN) that is assigned to a GameLift fleet resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:<region>::fleet/fleet-a1234567-b8c9-0d1e-2fa3-b45c6d7e8912.

httpStatus :: Int

The response's http status code.

newStartFleetActionsResponse

Arguments

:: Int
$sel:httpStatus:StartFleetActionsResponse'

-> StartFleetActionsResponse

Create a value of StartFleetActionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fleetId:StartFleetActions', startFleetActionsResponse_fleetId - A unique identifier for the fleet to restart actions on.

$sel:fleetArn:StartFleetActionsResponse', startFleetActionsResponse_fleetArn - The Amazon Resource Name (ARN) that is assigned to a GameLift fleet resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:<region>::fleet/fleet-a1234567-b8c9-0d1e-2fa3-b45c6d7e8912.

$sel:httpStatus:StartFleetActionsResponse', startFleetActionsResponse_httpStatus - The response's http status code.

Response Lenses

startFleetActionsResponse_fleetId :: Lens' StartFleetActionsResponse (Maybe Text)

A unique identifier for the fleet to restart actions on.

startFleetActionsResponse_fleetArn :: Lens' StartFleetActionsResponse (Maybe Text)

The Amazon Resource Name (ARN) that is assigned to a GameLift fleet resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:<region>::fleet/fleet-a1234567-b8c9-0d1e-2fa3-b45c6d7e8912.

startFleetActionsResponse_httpStatus :: Lens' StartFleetActionsResponse Int

The response's http status code.