Network.AWS.GameLift.StartGameSessionPlacement

Description

Places a request for a new game session in a queue (see CreateGameSessionQueue). When processing a placement request, Amazon GameLift searches for available resources on the queue's destinations, scanning each until it finds resources or the placement request times out.

A game session placement request can also request player sessions. When a new game session is successfully created, Amazon GameLift creates a player session for each player included in the request.

When placing a game session, by default Amazon GameLift tries each fleet in the order they are listed in the queue configuration. Ideally, a queue's destinations are listed in preference order.

Alternatively, when requesting a game session with players, you can also provide latency data for each player in relevant Regions. Latency data indicates the performance lag a player experiences when connected to a fleet in the Region. Amazon GameLift uses latency data to reorder the list of destinations to place the game session in a Region with minimal lag. If latency data is provided for multiple players, Amazon GameLift calculates each Region's average lag for all players and reorders to get the best game play across all players.

To place a new game session request, specify the following:

The queue name and a set of game session properties and settings

A unique ID (such as a UUID) for the placement. You use this ID to track the status of the placement request

(Optional) A set of player data and a unique player ID for each player that you are joining to the new game session (player data is optional, but if you include it, you must also provide a unique ID for each player)

Latency data for all players (if you want to optimize game play for the players)

If successful, a new game session placement is created.

To track the status of a placement request, call DescribeGameSessionPlacement and check the request's status. If the status is FULFILLED, a new game session has been created and a game session ARN and Region are referenced. If the placement request times out, you can resubmit the request or retry it with a different queue.

Related actions

CreateGameSession | DescribeGameSessions | DescribeGameSessionDetails | SearchGameSessions | UpdateGameSession | GetGameSessionLogUrl | StartGameSessionPlacement | DescribeGameSessionPlacement | StopGameSessionPlacement | All APIs by task

Creating a Request

data StartGameSessionPlacement

Represents the input for a request operation.

See: newStartGameSessionPlacement smart constructor.

Constructors

StartGameSessionPlacement'

Fields

gameProperties :: Maybe [GameProperty]

A set of custom properties for a game session, formatted as key:value pairs. These properties are passed to a game server process in the GameSession object with a request to start a new game session (see Start a Game Session).

gameSessionName :: Maybe Text

A descriptive label that is associated with a game session. Session names do not need to be unique.

playerLatencies :: Maybe [PlayerLatency]

A set of values, expressed in milliseconds, that indicates the amount of latency that a player experiences when connected to AWS Regions. This information is used to try to place the new game session where it can offer the best possible gameplay experience for the players.

gameSessionData :: Maybe Text

A set of custom game session properties, formatted as a single string value. This data is passed to a game server process in the GameSession object with a request to start a new game session (see Start a Game Session).

desiredPlayerSessions :: Maybe [DesiredPlayerSession]

Set of information on each player to create a player session for.

placementId :: Text

A unique identifier to assign to the new game session placement. This value is developer-defined. The value must be unique across all Regions and cannot be reused unless you are resubmitting a canceled or timed-out placement request.

gameSessionQueueName :: Text

Name of the queue to use to place the new game session. You can use either the queue name or ARN value.

maximumPlayerSessionCount :: Natural

The maximum number of players that can be connected simultaneously to the game session.

newStartGameSessionPlacement

Arguments

:: Text
$sel:placementId:StartGameSessionPlacement'

-> Text
$sel:gameSessionQueueName:StartGameSessionPlacement'

-> Natural
$sel:maximumPlayerSessionCount:StartGameSessionPlacement'

-> StartGameSessionPlacement

Create a value of StartGameSessionPlacement with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:gameProperties:StartGameSessionPlacement', startGameSessionPlacement_gameProperties - A set of custom properties for a game session, formatted as key:value pairs. These properties are passed to a game server process in the GameSession object with a request to start a new game session (see Start a Game Session).

$sel:gameSessionName:StartGameSessionPlacement', startGameSessionPlacement_gameSessionName - A descriptive label that is associated with a game session. Session names do not need to be unique.

$sel:playerLatencies:StartGameSessionPlacement', startGameSessionPlacement_playerLatencies - A set of values, expressed in milliseconds, that indicates the amount of latency that a player experiences when connected to AWS Regions. This information is used to try to place the new game session where it can offer the best possible gameplay experience for the players.

$sel:gameSessionData:StartGameSessionPlacement', startGameSessionPlacement_gameSessionData - A set of custom game session properties, formatted as a single string value. This data is passed to a game server process in the GameSession object with a request to start a new game session (see Start a Game Session).

$sel:desiredPlayerSessions:StartGameSessionPlacement', startGameSessionPlacement_desiredPlayerSessions - Set of information on each player to create a player session for.

$sel:placementId:StartGameSessionPlacement', startGameSessionPlacement_placementId - A unique identifier to assign to the new game session placement. This value is developer-defined. The value must be unique across all Regions and cannot be reused unless you are resubmitting a canceled or timed-out placement request.

$sel:gameSessionQueueName:StartGameSessionPlacement', startGameSessionPlacement_gameSessionQueueName - Name of the queue to use to place the new game session. You can use either the queue name or ARN value.

$sel:maximumPlayerSessionCount:StartGameSessionPlacement', startGameSessionPlacement_maximumPlayerSessionCount - The maximum number of players that can be connected simultaneously to the game session.

Request Lenses

startGameSessionPlacement_gameProperties :: Lens' StartGameSessionPlacement (Maybe [GameProperty])

A set of custom properties for a game session, formatted as key:value pairs. These properties are passed to a game server process in the GameSession object with a request to start a new game session (see Start a Game Session).

startGameSessionPlacement_gameSessionName :: Lens' StartGameSessionPlacement (Maybe Text)

A descriptive label that is associated with a game session. Session names do not need to be unique.

startGameSessionPlacement_playerLatencies :: Lens' StartGameSessionPlacement (Maybe [PlayerLatency])

A set of values, expressed in milliseconds, that indicates the amount of latency that a player experiences when connected to AWS Regions. This information is used to try to place the new game session where it can offer the best possible gameplay experience for the players.

startGameSessionPlacement_gameSessionData :: Lens' StartGameSessionPlacement (Maybe Text)

A set of custom game session properties, formatted as a single string value. This data is passed to a game server process in the GameSession object with a request to start a new game session (see Start a Game Session).

startGameSessionPlacement_desiredPlayerSessions :: Lens' StartGameSessionPlacement (Maybe [DesiredPlayerSession])

Set of information on each player to create a player session for.

startGameSessionPlacement_placementId :: Lens' StartGameSessionPlacement Text

A unique identifier to assign to the new game session placement. This value is developer-defined. The value must be unique across all Regions and cannot be reused unless you are resubmitting a canceled or timed-out placement request.

startGameSessionPlacement_gameSessionQueueName :: Lens' StartGameSessionPlacement Text

Name of the queue to use to place the new game session. You can use either the queue name or ARN value.

startGameSessionPlacement_maximumPlayerSessionCount :: Lens' StartGameSessionPlacement Natural

The maximum number of players that can be connected simultaneously to the game session.

Destructuring the Response

data StartGameSessionPlacementResponse

Represents the returned data in response to a request operation.

See: newStartGameSessionPlacementResponse smart constructor.

Constructors

StartGameSessionPlacementResponse'

Fields

gameSessionPlacement :: Maybe GameSessionPlacement

Object that describes the newly created game session placement. This object includes all the information provided in the request, as well as start/end time stamps and placement status.

httpStatus :: Int

The response's http status code.

newStartGameSessionPlacementResponse

Arguments

:: Int
$sel:httpStatus:StartGameSessionPlacementResponse'

-> StartGameSessionPlacementResponse

Create a value of StartGameSessionPlacementResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:gameSessionPlacement:StartGameSessionPlacementResponse', startGameSessionPlacementResponse_gameSessionPlacement - Object that describes the newly created game session placement. This object includes all the information provided in the request, as well as start/end time stamps and placement status.

$sel:httpStatus:StartGameSessionPlacementResponse', startGameSessionPlacementResponse_httpStatus - The response's http status code.

Response Lenses

startGameSessionPlacementResponse_gameSessionPlacement :: Lens' StartGameSessionPlacementResponse (Maybe GameSessionPlacement)

Object that describes the newly created game session placement. This object includes all the information provided in the request, as well as start/end time stamps and placement status.

startGameSessionPlacementResponse_httpStatus :: Lens' StartGameSessionPlacementResponse Int

The response's http status code.