Network.AWS.GameLift.StartMatchBackfill

Description

Finds new players to fill open slots in currently running game sessions. The backfill match process is essentially identical to the process of forming new matches. Backfill requests use the same matchmaker that was used to make the original match, and they provide matchmaking data for all players currently in the game session. FlexMatch uses this information to select new players so that backfilled match continues to meet the original match requirements.

When using FlexMatch with GameLift managed hosting, you can request a backfill match from a client service by calling this operation with a GameSession identifier. You also have the option of making backfill requests directly from your game server. In response to a request, FlexMatch creates player sessions for the new players, updates the GameSession resource, and sends updated matchmaking data to the game server. You can request a backfill match at any point after a game session is started. Each game session can have only one active backfill request at a time; a subsequent request automatically replaces the earlier request.

When using FlexMatch as a standalone component, request a backfill match by calling this operation without a game session identifier. As with newly formed matches, matchmaking results are returned in a matchmaking event so that your game can update the game session that is being backfilled.

To request a backfill match, specify a unique ticket ID, the original matchmaking configuration, and matchmaking data for all current players in the game session being backfilled. Optionally, specify the GameSession ARN. If successful, a match backfill ticket is created and returned with status set to QUEUED. Track the status of backfill tickets using the same method for tracking tickets for new matches.

Learn more

Backfill existing games with FlexMatch

Matchmaking events (reference)

How GameLift FlexMatch works

Related actions

StartMatchmaking | DescribeMatchmaking | StopMatchmaking | AcceptMatch | StartMatchBackfill | All APIs by task

Creating a Request

data StartMatchBackfill

Represents the input for a request operation.

See: newStartMatchBackfill smart constructor.

Constructors

StartMatchBackfill'

Fields

ticketId :: Maybe Text

A unique identifier for a matchmaking ticket. If no ticket ID is specified here, Amazon GameLift will generate one in the form of a UUID. Use this identifier to track the match backfill ticket status and retrieve match results.

gameSessionArn :: Maybe Text

A unique identifier for the game session. Use the game session ID. When using FlexMatch as a standalone matchmaking solution, this parameter is not needed.

configurationName :: Text

Name of the matchmaker to use for this request. You can use either the configuration name or ARN value. The ARN of the matchmaker that was used with the original game session is listed in the GameSession object, MatchmakerData property.

players :: [Player]

Match information on all players that are currently assigned to the game session. This information is used by the matchmaker to find new players and add them to the existing game.

PlayerID, PlayerAttributes, Team -- This information is maintained in the GameSession object, MatchmakerData property, for all players who are currently assigned to the game session. The matchmaker data is in JSON syntax, formatted as a string. For more details, see Match Data.

LatencyInMs -- If the matchmaker uses player latency, include a latency value, in milliseconds, for the Region that the game session is currently in. Do not include latency values for any other Region.

newStartMatchBackfill

Arguments

:: Text
$sel:configurationName:StartMatchBackfill'

-> StartMatchBackfill

Create a value of StartMatchBackfill with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ticketId:StartMatchBackfill', startMatchBackfill_ticketId - A unique identifier for a matchmaking ticket. If no ticket ID is specified here, Amazon GameLift will generate one in the form of a UUID. Use this identifier to track the match backfill ticket status and retrieve match results.

$sel:gameSessionArn:StartMatchBackfill', startMatchBackfill_gameSessionArn - A unique identifier for the game session. Use the game session ID. When using FlexMatch as a standalone matchmaking solution, this parameter is not needed.

$sel:configurationName:StartMatchBackfill', startMatchBackfill_configurationName - Name of the matchmaker to use for this request. You can use either the configuration name or ARN value. The ARN of the matchmaker that was used with the original game session is listed in the GameSession object, MatchmakerData property.

$sel:players:StartMatchBackfill', startMatchBackfill_players - Match information on all players that are currently assigned to the game session. This information is used by the matchmaker to find new players and add them to the existing game.

PlayerID, PlayerAttributes, Team -- This information is maintained in the GameSession object, MatchmakerData property, for all players who are currently assigned to the game session. The matchmaker data is in JSON syntax, formatted as a string. For more details, see Match Data.

LatencyInMs -- If the matchmaker uses player latency, include a latency value, in milliseconds, for the Region that the game session is currently in. Do not include latency values for any other Region.

Request Lenses

startMatchBackfill_ticketId :: Lens' StartMatchBackfill (Maybe Text)

A unique identifier for a matchmaking ticket. If no ticket ID is specified here, Amazon GameLift will generate one in the form of a UUID. Use this identifier to track the match backfill ticket status and retrieve match results.

startMatchBackfill_gameSessionArn :: Lens' StartMatchBackfill (Maybe Text)

A unique identifier for the game session. Use the game session ID. When using FlexMatch as a standalone matchmaking solution, this parameter is not needed.

startMatchBackfill_configurationName :: Lens' StartMatchBackfill Text

Name of the matchmaker to use for this request. You can use either the configuration name or ARN value. The ARN of the matchmaker that was used with the original game session is listed in the GameSession object, MatchmakerData property.

startMatchBackfill_players :: Lens' StartMatchBackfill [Player]

Match information on all players that are currently assigned to the game session. This information is used by the matchmaker to find new players and add them to the existing game.

PlayerID, PlayerAttributes, Team -- This information is maintained in the GameSession object, MatchmakerData property, for all players who are currently assigned to the game session. The matchmaker data is in JSON syntax, formatted as a string. For more details, see Match Data.

LatencyInMs -- If the matchmaker uses player latency, include a latency value, in milliseconds, for the Region that the game session is currently in. Do not include latency values for any other Region.

Destructuring the Response

data StartMatchBackfillResponse

Represents the returned data in response to a request operation.

See: newStartMatchBackfillResponse smart constructor.

Constructors

StartMatchBackfillResponse'

Fields

matchmakingTicket :: Maybe MatchmakingTicket

Ticket representing the backfill matchmaking request. This object includes the information in the request, ticket status, and match results as generated during the matchmaking process.

httpStatus :: Int

The response's http status code.

newStartMatchBackfillResponse

Arguments

:: Int
$sel:httpStatus:StartMatchBackfillResponse'

-> StartMatchBackfillResponse

Create a value of StartMatchBackfillResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:matchmakingTicket:StartMatchBackfillResponse', startMatchBackfillResponse_matchmakingTicket - Ticket representing the backfill matchmaking request. This object includes the information in the request, ticket status, and match results as generated during the matchmaking process.

$sel:httpStatus:StartMatchBackfillResponse', startMatchBackfillResponse_httpStatus - The response's http status code.

Response Lenses

startMatchBackfillResponse_matchmakingTicket :: Lens' StartMatchBackfillResponse (Maybe MatchmakingTicket)

Ticket representing the backfill matchmaking request. This object includes the information in the request, ticket status, and match results as generated during the matchmaking process.

startMatchBackfillResponse_httpStatus :: Lens' StartMatchBackfillResponse Int

The response's http status code.