Network.AWS.GameLift.StopMatchmaking

Description

Cancels a matchmaking ticket or match backfill ticket that is currently being processed. To stop the matchmaking operation, specify the ticket ID. If successful, work on the ticket is stopped, and the ticket status is changed to CANCELLED.

This call is also used to turn off automatic backfill for an individual game session. This is for game sessions that are created with a matchmaking configuration that has automatic backfill enabled. The ticket ID is included in the MatchmakerData of an updated game session object, which is provided to the game server.

If the operation is successful, the service sends back an empty JSON struct with the HTTP 200 response (not an empty HTTP body).

Learn more

Add FlexMatch to a game client

Related actions

StartMatchmaking | DescribeMatchmaking | StopMatchmaking | AcceptMatch | StartMatchBackfill | All APIs by task

Creating a Request

data StopMatchmaking

Represents the input for a request operation.

See: newStopMatchmaking smart constructor.

Constructors

StopMatchmaking'

Fields

ticketId :: Text

A unique identifier for a matchmaking ticket.

newStopMatchmaking

Arguments

:: Text
$sel:ticketId:StopMatchmaking'

-> StopMatchmaking

Create a value of StopMatchmaking with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ticketId:StopMatchmaking', stopMatchmaking_ticketId - A unique identifier for a matchmaking ticket.

Request Lenses

stopMatchmaking_ticketId :: Lens' StopMatchmaking Text

A unique identifier for a matchmaking ticket.

Destructuring the Response

data StopMatchmakingResponse

See: newStopMatchmakingResponse smart constructor.

Constructors

StopMatchmakingResponse'

Fields

httpStatus :: Int

The response's http status code.

newStopMatchmakingResponse

Arguments

:: Int
$sel:httpStatus:StopMatchmakingResponse'

-> StopMatchmakingResponse

Create a value of StopMatchmakingResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:StopMatchmakingResponse', stopMatchmakingResponse_httpStatus - The response's http status code.

Response Lenses

stopMatchmakingResponse_httpStatus :: Lens' StopMatchmakingResponse Int

The response's http status code.