Network.AWS.GameLift.Types.Alias

Description

Documentation

data Alias

Properties that describe an alias resource.

Related actions

CreateAlias | ListAliases | DescribeAlias | UpdateAlias | DeleteAlias | ResolveAlias | All APIs by task

See: newAlias smart constructor.

Constructors

Alias'

Fields

routingStrategy :: Maybe RoutingStrategy

The routing configuration, including routing type and fleet target, for the alias.

creationTime :: Maybe POSIX

A time stamp indicating when this data object was created. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

name :: Maybe Text

A descriptive label that is associated with an alias. Alias names do not need to be unique.

description :: Maybe Text

A human-readable description of an alias.

aliasArn :: Maybe Text

The Amazon Resource Name (ARN) that is assigned to a GameLift alias resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:<region>::alias/alias-a1234567-b8c9-0d1e-2fa3-b45c6d7e8912. In a GameLift alias ARN, the resource ID matches the alias ID value.

aliasId :: Maybe Text

A unique identifier for the alias. Alias IDs are unique within a Region.

lastUpdatedTime :: Maybe POSIX

The time that this data object was last modified. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

newAlias :: Alias

Create a value of Alias with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:routingStrategy:Alias', alias_routingStrategy - The routing configuration, including routing type and fleet target, for the alias.

$sel:creationTime:Alias', alias_creationTime - A time stamp indicating when this data object was created. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

$sel:name:Alias', alias_name - A descriptive label that is associated with an alias. Alias names do not need to be unique.

$sel:description:Alias', alias_description - A human-readable description of an alias.

$sel:aliasArn:Alias', alias_aliasArn - The Amazon Resource Name (ARN) that is assigned to a GameLift alias resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:<region>::alias/alias-a1234567-b8c9-0d1e-2fa3-b45c6d7e8912. In a GameLift alias ARN, the resource ID matches the alias ID value.

$sel:aliasId:Alias', alias_aliasId - A unique identifier for the alias. Alias IDs are unique within a Region.

$sel:lastUpdatedTime:Alias', alias_lastUpdatedTime - The time that this data object was last modified. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

alias_routingStrategy :: Lens' Alias (Maybe RoutingStrategy)

The routing configuration, including routing type and fleet target, for the alias.

alias_creationTime :: Lens' Alias (Maybe UTCTime)

A time stamp indicating when this data object was created. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

alias_name :: Lens' Alias (Maybe Text)

A descriptive label that is associated with an alias. Alias names do not need to be unique.

alias_description :: Lens' Alias (Maybe Text)

A human-readable description of an alias.

alias_aliasArn :: Lens' Alias (Maybe Text)

The Amazon Resource Name (ARN) that is assigned to a GameLift alias resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:<region>::alias/alias-a1234567-b8c9-0d1e-2fa3-b45c6d7e8912. In a GameLift alias ARN, the resource ID matches the alias ID value.

alias_aliasId :: Lens' Alias (Maybe Text)

A unique identifier for the alias. Alias IDs are unique within a Region.

alias_lastUpdatedTime :: Lens' Alias (Maybe UTCTime)

The time that this data object was last modified. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").