Network.AWS.GameLift.Types.CertificateConfiguration

Description

Documentation

data CertificateConfiguration

Determines whether a TLS/SSL certificate is generated for a fleet. This feature must be enabled when creating the fleet. All instances in a fleet share the same certificate. The certificate can be retrieved by calling the GameLift Server SDK operation GetInstanceCertificate.

A fleet's certificate configuration is part of FleetAttributes.

See: newCertificateConfiguration smart constructor.

Constructors

CertificateConfiguration'

Fields

certificateType :: CertificateType

Indicates whether a TLS/SSL certificate is generated for a fleet.

Valid values include:

GENERATED - Generate a TLS/SSL certificate for this fleet.

DISABLED - (default) Do not generate a TLS/SSL certificate for this fleet.

newCertificateConfiguration

Arguments

:: CertificateType
$sel:certificateType:CertificateConfiguration'

-> CertificateConfiguration

Create a value of CertificateConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:certificateType:CertificateConfiguration', certificateConfiguration_certificateType - Indicates whether a TLS/SSL certificate is generated for a fleet.

Valid values include:

GENERATED - Generate a TLS/SSL certificate for this fleet.

DISABLED - (default) Do not generate a TLS/SSL certificate for this fleet.

certificateConfiguration_certificateType :: Lens' CertificateConfiguration CertificateType

Indicates whether a TLS/SSL certificate is generated for a fleet.

Valid values include:

GENERATED - Generate a TLS/SSL certificate for this fleet.

DISABLED - (default) Do not generate a TLS/SSL certificate for this fleet.