Network.AWS.GameLift.Types.EC2InstanceCounts

Description

Documentation

data EC2InstanceCounts

Resource capacity settings. Fleet capacity is measured in EC2 instances. Pending and terminating counts are non-zero when the fleet capacity is adjusting to a scaling event or if access to resources is temporarily affected.

EC2 instance counts are part of FleetCapacity.

See: newEC2InstanceCounts smart constructor.

Constructors

EC2InstanceCounts'

Fields

idle :: Maybe Natural

Number of active instances that are not currently hosting a game session.

terminating :: Maybe Natural

Number of instances that are no longer active but haven't yet been terminated.

pending :: Maybe Natural

Number of instances that are starting but not yet active.

maximum :: Maybe Natural

The maximum instance count value allowed.

desired :: Maybe Natural

Ideal number of active instances. GameLift will always try to maintain the desired number of instances. Capacity is scaled up or down by changing the desired instances.

minimum :: Maybe Natural

The minimum instance count value allowed.

active :: Maybe Natural

Actual number of instances that are ready to host game sessions.

newEC2InstanceCounts :: EC2InstanceCounts

Create a value of EC2InstanceCounts with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:idle:EC2InstanceCounts', eC2InstanceCounts_idle - Number of active instances that are not currently hosting a game session.

$sel:terminating:EC2InstanceCounts', eC2InstanceCounts_terminating - Number of instances that are no longer active but haven't yet been terminated.

$sel:pending:EC2InstanceCounts', eC2InstanceCounts_pending - Number of instances that are starting but not yet active.

$sel:maximum:EC2InstanceCounts', eC2InstanceCounts_maximum - The maximum instance count value allowed.

$sel:desired:EC2InstanceCounts', eC2InstanceCounts_desired - Ideal number of active instances. GameLift will always try to maintain the desired number of instances. Capacity is scaled up or down by changing the desired instances.

$sel:minimum:EC2InstanceCounts', eC2InstanceCounts_minimum - The minimum instance count value allowed.

$sel:active:EC2InstanceCounts', eC2InstanceCounts_active - Actual number of instances that are ready to host game sessions.

eC2InstanceCounts_idle :: Lens' EC2InstanceCounts (Maybe Natural)

Number of active instances that are not currently hosting a game session.

eC2InstanceCounts_terminating :: Lens' EC2InstanceCounts (Maybe Natural)

Number of instances that are no longer active but haven't yet been terminated.

eC2InstanceCounts_pending :: Lens' EC2InstanceCounts (Maybe Natural)

Number of instances that are starting but not yet active.

eC2InstanceCounts_maximum :: Lens' EC2InstanceCounts (Maybe Natural)

The maximum instance count value allowed.

eC2InstanceCounts_desired :: Lens' EC2InstanceCounts (Maybe Natural)

Ideal number of active instances. GameLift will always try to maintain the desired number of instances. Capacity is scaled up or down by changing the desired instances.

eC2InstanceCounts_minimum :: Lens' EC2InstanceCounts (Maybe Natural)

The minimum instance count value allowed.

eC2InstanceCounts_active :: Lens' EC2InstanceCounts (Maybe Natural)

Actual number of instances that are ready to host game sessions.