Network.AWS.GameLift.Types.EC2InstanceLimit

Description

Documentation

data EC2InstanceLimit

The GameLift service limits for an EC2 instance type and current utilization. GameLift allows AWS accounts a maximum number of instances, per instance type, per AWS Region or location, for use with GameLift. You can request an limit increase for your account by using the Service limits page in the GameLift console.

Related actions

DescribeEC2InstanceLimits

See: newEC2InstanceLimit smart constructor.

Constructors

EC2InstanceLimit'

Fields

instanceLimit :: Maybe Natural

The number of instances that is allowed for the specified instance type and location.

currentInstances :: Maybe Natural

The number of instances for the specified type and location that are currently being used by the AWS account.

eC2InstanceType :: Maybe EC2InstanceType

The name of an EC2 instance type. See Amazon EC2 Instance Types for detailed descriptions.

location :: Maybe Text

An AWS Region code, such as us-west-2.

newEC2InstanceLimit :: EC2InstanceLimit

Create a value of EC2InstanceLimit with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceLimit:EC2InstanceLimit', eC2InstanceLimit_instanceLimit - The number of instances that is allowed for the specified instance type and location.

$sel:currentInstances:EC2InstanceLimit', eC2InstanceLimit_currentInstances - The number of instances for the specified type and location that are currently being used by the AWS account.

$sel:eC2InstanceType:EC2InstanceLimit', eC2InstanceLimit_eC2InstanceType - The name of an EC2 instance type. See Amazon EC2 Instance Types for detailed descriptions.

$sel:location:EC2InstanceLimit', eC2InstanceLimit_location - An AWS Region code, such as us-west-2.

eC2InstanceLimit_instanceLimit :: Lens' EC2InstanceLimit (Maybe Natural)

The number of instances that is allowed for the specified instance type and location.

eC2InstanceLimit_currentInstances :: Lens' EC2InstanceLimit (Maybe Natural)

The number of instances for the specified type and location that are currently being used by the AWS account.

eC2InstanceLimit_eC2InstanceType :: Lens' EC2InstanceLimit (Maybe EC2InstanceType)

The name of an EC2 instance type. See Amazon EC2 Instance Types for detailed descriptions.

eC2InstanceLimit_location :: Lens' EC2InstanceLimit (Maybe Text)

An AWS Region code, such as us-west-2.