Network.AWS.GameLift.Types.FleetAttributes

Description

Documentation

data FleetAttributes

Describes a GameLift fleet of game hosting resources.

Related actions

CreateFleet | DescribeFleetAttributes

See: newFleetAttributes smart constructor.

Constructors

FleetAttributes'

Fields

creationTime :: Maybe POSIX

A time stamp indicating when this data object was created. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

status :: Maybe FleetStatus

Current status of the fleet. Possible fleet statuses include the following:

NEW -- A new fleet has been defined and desired instances is set to 1.

DOWNLOADING/VALIDATING/BUILDING/ACTIVATING -- GameLift is setting up the new fleet, creating new instances with the game build or Realtime script and starting server processes.

ACTIVE -- Hosts can now accept game sessions.

ERROR -- An error occurred when downloading, validating, building, or activating the fleet.

DELETING -- Hosts are responding to a delete fleet request.

TERMINATED -- The fleet no longer exists.

instanceType :: Maybe EC2InstanceType

The EC2 instance type that determines the computing resources of each instance in the fleet. Instance type defines the CPU, memory, storage, and networking capacity. See Amazon EC2 Instance Types for detailed descriptions.

fleetType :: Maybe FleetType

The kind of instances, On-Demand or Spot, that this fleet uses.

fleetId :: Maybe Text

A unique identifier for the fleet.

fleetArn :: Maybe Text

The Amazon Resource Name (ARN) that is assigned to a GameLift fleet resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:<region>::fleet/fleet-a1234567-b8c9-0d1e-2fa3-b45c6d7e8912. In a GameLift fleet ARN, the resource ID matches the FleetId value.

instanceRoleArn :: Maybe Text

A unique identifier for an AWS IAM role that manages access to your AWS services. With an instance role ARN set, any application that runs on an instance in this fleet can assume the role, including install scripts, server processes, and daemons (background processes). Create a role or look up a role's ARN by using the IAM dashboard in the AWS Management Console. Learn more about using on-box credentials for your game servers at Access external resources from a game server.

certificateConfiguration :: Maybe CertificateConfiguration

Indicates whether a TLS/SSL certificate was generated for the fleet.

serverLaunchPath :: Maybe Text

This parameter is no longer used. Server launch paths are now defined using the fleet's RuntimeConfiguration parameter. Requests that use this parameter instead continue to be valid.

logPaths :: Maybe [Text]

This parameter is no longer used. Game session log paths are now defined using the GameLift server API ProcessReady() logParameters. See more information in the Server API Reference.

serverLaunchParameters :: Maybe Text

This parameter is no longer used. Server launch parameters are now defined using the fleet's RuntimeConfiguration parameter. Requests that use this parameter instead continue to be valid.

scriptArn :: Maybe Text

The Amazon Resource Name (ARN) associated with the GameLift script resource that is deployed on instances in this fleet. In a GameLift script ARN, the resource ID matches the ScriptId value.

name :: Maybe Text

A descriptive label that is associated with a fleet. Fleet names do not need to be unique.

newGameSessionProtectionPolicy' :: Maybe ProtectionPolicy

The type of game session protection to set on all new instances that are started in the fleet.

NoProtection -- The game session can be terminated during a scale-down event.

FullProtection -- If the game session is in an ACTIVE status, it cannot be terminated during a scale-down event.

stoppedActions :: Maybe (NonEmpty FleetAction)

A list of fleet activity that has been suspended using StopFleetActions. This includes fleet auto-scaling.

terminationTime :: Maybe POSIX

A time stamp indicating when this data object was terminated. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

resourceCreationLimitPolicy :: Maybe ResourceCreationLimitPolicy

The fleet policy that limits the number of game sessions an individual player can create over a span of time.

description :: Maybe Text

A human-readable description of the fleet.

buildId :: Maybe Text

A unique identifier for the build resource that is deployed on instances in this fleet.

buildArn :: Maybe Text

The Amazon Resource Name (ARN) associated with the GameLift build resource that is deployed on instances in this fleet. In a GameLift build ARN, the resource ID matches the BuildId value.

metricGroups :: Maybe [Text]

Name of a metric group that metrics for this fleet are added to. In Amazon CloudWatch, you can view aggregated metrics for fleets that are in a metric group. A fleet can be included in only one metric group at a time.

operatingSystem :: Maybe OperatingSystem

The operating system of the fleet's computing resources. A fleet's operating system is determined by the OS of the build or script that is deployed on this fleet.

scriptId :: Maybe Text

A unique identifier for the Realtime script resource that is deployed on instances in this fleet.

newFleetAttributes :: FleetAttributes

Create a value of FleetAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:FleetAttributes', fleetAttributes_creationTime - A time stamp indicating when this data object was created. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

$sel:status:FleetAttributes', fleetAttributes_status - Current status of the fleet. Possible fleet statuses include the following:

NEW -- A new fleet has been defined and desired instances is set to 1.

DOWNLOADING/VALIDATING/BUILDING/ACTIVATING -- GameLift is setting up the new fleet, creating new instances with the game build or Realtime script and starting server processes.

ACTIVE -- Hosts can now accept game sessions.

ERROR -- An error occurred when downloading, validating, building, or activating the fleet.

DELETING -- Hosts are responding to a delete fleet request.

TERMINATED -- The fleet no longer exists.

$sel:instanceType:FleetAttributes', fleetAttributes_instanceType - The EC2 instance type that determines the computing resources of each instance in the fleet. Instance type defines the CPU, memory, storage, and networking capacity. See Amazon EC2 Instance Types for detailed descriptions.

$sel:fleetType:FleetAttributes', fleetAttributes_fleetType - The kind of instances, On-Demand or Spot, that this fleet uses.

$sel:fleetId:FleetAttributes', fleetAttributes_fleetId - A unique identifier for the fleet.

$sel:fleetArn:FleetAttributes', fleetAttributes_fleetArn - The Amazon Resource Name (ARN) that is assigned to a GameLift fleet resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:<region>::fleet/fleet-a1234567-b8c9-0d1e-2fa3-b45c6d7e8912. In a GameLift fleet ARN, the resource ID matches the FleetId value.

$sel:instanceRoleArn:FleetAttributes', fleetAttributes_instanceRoleArn - A unique identifier for an AWS IAM role that manages access to your AWS services. With an instance role ARN set, any application that runs on an instance in this fleet can assume the role, including install scripts, server processes, and daemons (background processes). Create a role or look up a role's ARN by using the IAM dashboard in the AWS Management Console. Learn more about using on-box credentials for your game servers at Access external resources from a game server.

$sel:certificateConfiguration:FleetAttributes', fleetAttributes_certificateConfiguration - Indicates whether a TLS/SSL certificate was generated for the fleet.

$sel:serverLaunchPath:FleetAttributes', fleetAttributes_serverLaunchPath - This parameter is no longer used. Server launch paths are now defined using the fleet's RuntimeConfiguration parameter. Requests that use this parameter instead continue to be valid.

$sel:logPaths:FleetAttributes', fleetAttributes_logPaths - This parameter is no longer used. Game session log paths are now defined using the GameLift server API ProcessReady() logParameters. See more information in the Server API Reference.

$sel:serverLaunchParameters:FleetAttributes', fleetAttributes_serverLaunchParameters - This parameter is no longer used. Server launch parameters are now defined using the fleet's RuntimeConfiguration parameter. Requests that use this parameter instead continue to be valid.

$sel:scriptArn:FleetAttributes', fleetAttributes_scriptArn - The Amazon Resource Name (ARN) associated with the GameLift script resource that is deployed on instances in this fleet. In a GameLift script ARN, the resource ID matches the ScriptId value.

$sel:name:FleetAttributes', fleetAttributes_name - A descriptive label that is associated with a fleet. Fleet names do not need to be unique.

$sel:newGameSessionProtectionPolicy':FleetAttributes', fleetAttributes_newGameSessionProtectionPolicy - The type of game session protection to set on all new instances that are started in the fleet.

NoProtection -- The game session can be terminated during a scale-down event.

FullProtection -- If the game session is in an ACTIVE status, it cannot be terminated during a scale-down event.

$sel:stoppedActions:FleetAttributes', fleetAttributes_stoppedActions - A list of fleet activity that has been suspended using StopFleetActions. This includes fleet auto-scaling.

$sel:terminationTime:FleetAttributes', fleetAttributes_terminationTime - A time stamp indicating when this data object was terminated. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

$sel:resourceCreationLimitPolicy:FleetAttributes', fleetAttributes_resourceCreationLimitPolicy - The fleet policy that limits the number of game sessions an individual player can create over a span of time.

$sel:description:FleetAttributes', fleetAttributes_description - A human-readable description of the fleet.

$sel:buildId:FleetAttributes', fleetAttributes_buildId - A unique identifier for the build resource that is deployed on instances in this fleet.

$sel:buildArn:FleetAttributes', fleetAttributes_buildArn - The Amazon Resource Name (ARN) associated with the GameLift build resource that is deployed on instances in this fleet. In a GameLift build ARN, the resource ID matches the BuildId value.

$sel:metricGroups:FleetAttributes', fleetAttributes_metricGroups - Name of a metric group that metrics for this fleet are added to. In Amazon CloudWatch, you can view aggregated metrics for fleets that are in a metric group. A fleet can be included in only one metric group at a time.

$sel:operatingSystem:FleetAttributes', fleetAttributes_operatingSystem - The operating system of the fleet's computing resources. A fleet's operating system is determined by the OS of the build or script that is deployed on this fleet.

$sel:scriptId:FleetAttributes', fleetAttributes_scriptId - A unique identifier for the Realtime script resource that is deployed on instances in this fleet.

fleetAttributes_creationTime :: Lens' FleetAttributes (Maybe UTCTime)

A time stamp indicating when this data object was created. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

fleetAttributes_status :: Lens' FleetAttributes (Maybe FleetStatus)

Current status of the fleet. Possible fleet statuses include the following:

NEW -- A new fleet has been defined and desired instances is set to 1.

DOWNLOADING/VALIDATING/BUILDING/ACTIVATING -- GameLift is setting up the new fleet, creating new instances with the game build or Realtime script and starting server processes.

ACTIVE -- Hosts can now accept game sessions.

ERROR -- An error occurred when downloading, validating, building, or activating the fleet.

DELETING -- Hosts are responding to a delete fleet request.

TERMINATED -- The fleet no longer exists.

fleetAttributes_instanceType :: Lens' FleetAttributes (Maybe EC2InstanceType)

The EC2 instance type that determines the computing resources of each instance in the fleet. Instance type defines the CPU, memory, storage, and networking capacity. See Amazon EC2 Instance Types for detailed descriptions.

fleetAttributes_fleetType :: Lens' FleetAttributes (Maybe FleetType)

The kind of instances, On-Demand or Spot, that this fleet uses.

fleetAttributes_fleetId :: Lens' FleetAttributes (Maybe Text)

A unique identifier for the fleet.

fleetAttributes_fleetArn :: Lens' FleetAttributes (Maybe Text)

The Amazon Resource Name (ARN) that is assigned to a GameLift fleet resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:<region>::fleet/fleet-a1234567-b8c9-0d1e-2fa3-b45c6d7e8912. In a GameLift fleet ARN, the resource ID matches the FleetId value.

fleetAttributes_instanceRoleArn :: Lens' FleetAttributes (Maybe Text)

A unique identifier for an AWS IAM role that manages access to your AWS services. With an instance role ARN set, any application that runs on an instance in this fleet can assume the role, including install scripts, server processes, and daemons (background processes). Create a role or look up a role's ARN by using the IAM dashboard in the AWS Management Console. Learn more about using on-box credentials for your game servers at Access external resources from a game server.

fleetAttributes_certificateConfiguration :: Lens' FleetAttributes (Maybe CertificateConfiguration)

Indicates whether a TLS/SSL certificate was generated for the fleet.

fleetAttributes_serverLaunchPath :: Lens' FleetAttributes (Maybe Text)

This parameter is no longer used. Server launch paths are now defined using the fleet's RuntimeConfiguration parameter. Requests that use this parameter instead continue to be valid.

fleetAttributes_logPaths :: Lens' FleetAttributes (Maybe [Text])

This parameter is no longer used. Game session log paths are now defined using the GameLift server API ProcessReady() logParameters. See more information in the Server API Reference.

fleetAttributes_serverLaunchParameters :: Lens' FleetAttributes (Maybe Text)

This parameter is no longer used. Server launch parameters are now defined using the fleet's RuntimeConfiguration parameter. Requests that use this parameter instead continue to be valid.

fleetAttributes_scriptArn :: Lens' FleetAttributes (Maybe Text)

The Amazon Resource Name (ARN) associated with the GameLift script resource that is deployed on instances in this fleet. In a GameLift script ARN, the resource ID matches the ScriptId value.

fleetAttributes_name :: Lens' FleetAttributes (Maybe Text)

A descriptive label that is associated with a fleet. Fleet names do not need to be unique.

fleetAttributes_newGameSessionProtectionPolicy :: Lens' FleetAttributes (Maybe ProtectionPolicy)

The type of game session protection to set on all new instances that are started in the fleet.

NoProtection -- The game session can be terminated during a scale-down event.

FullProtection -- If the game session is in an ACTIVE status, it cannot be terminated during a scale-down event.

fleetAttributes_stoppedActions :: Lens' FleetAttributes (Maybe (NonEmpty FleetAction))

A list of fleet activity that has been suspended using StopFleetActions. This includes fleet auto-scaling.

fleetAttributes_terminationTime :: Lens' FleetAttributes (Maybe UTCTime)

A time stamp indicating when this data object was terminated. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

fleetAttributes_resourceCreationLimitPolicy :: Lens' FleetAttributes (Maybe ResourceCreationLimitPolicy)

The fleet policy that limits the number of game sessions an individual player can create over a span of time.

fleetAttributes_description :: Lens' FleetAttributes (Maybe Text)

A human-readable description of the fleet.

fleetAttributes_buildId :: Lens' FleetAttributes (Maybe Text)

A unique identifier for the build resource that is deployed on instances in this fleet.

fleetAttributes_buildArn :: Lens' FleetAttributes (Maybe Text)

The Amazon Resource Name (ARN) associated with the GameLift build resource that is deployed on instances in this fleet. In a GameLift build ARN, the resource ID matches the BuildId value.

fleetAttributes_metricGroups :: Lens' FleetAttributes (Maybe [Text])

Name of a metric group that metrics for this fleet are added to. In Amazon CloudWatch, you can view aggregated metrics for fleets that are in a metric group. A fleet can be included in only one metric group at a time.

fleetAttributes_operatingSystem :: Lens' FleetAttributes (Maybe OperatingSystem)

The operating system of the fleet's computing resources. A fleet's operating system is determined by the OS of the build or script that is deployed on this fleet.

fleetAttributes_scriptId :: Lens' FleetAttributes (Maybe Text)

A unique identifier for the Realtime script resource that is deployed on instances in this fleet.