Network.AWS.GameLift.Types.FleetCapacity

Description

Documentation

data FleetCapacity

Current resource capacity settings in a specified fleet or location. The location value might refer to a fleet's remote location or its home Region.

Related actions

DescribeFleetCapacity | DescribeFleetLocationCapacity | UpdateFleetCapacity

See: newFleetCapacity smart constructor.

Constructors

FleetCapacity'

Fields

instanceType :: Maybe EC2InstanceType

The EC2 instance type that is used for all instances in a fleet. The instance type determines the computing resources in use, including CPU, memory, storage, and networking capacity. See Amazon EC2 Instance Types for detailed descriptions.

fleetId :: Maybe Text

A unique identifier for the fleet associated with the location.

fleetArn :: Maybe Text

The Amazon Resource Name (ARN) that is assigned to a GameLift fleet resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:<region>::fleet/fleet-a1234567-b8c9-0d1e-2fa3-b45c6d7e8912.

instanceCounts :: Maybe EC2InstanceCounts

The current instance count and capacity settings for the fleet location.

location :: Maybe Text

The fleet location for the instance count information, expressed as an AWS Region code, such as us-west-2.

newFleetCapacity :: FleetCapacity

Create a value of FleetCapacity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceType:FleetCapacity', fleetCapacity_instanceType - The EC2 instance type that is used for all instances in a fleet. The instance type determines the computing resources in use, including CPU, memory, storage, and networking capacity. See Amazon EC2 Instance Types for detailed descriptions.

$sel:fleetId:FleetCapacity', fleetCapacity_fleetId - A unique identifier for the fleet associated with the location.

$sel:fleetArn:FleetCapacity', fleetCapacity_fleetArn - The Amazon Resource Name (ARN) that is assigned to a GameLift fleet resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:<region>::fleet/fleet-a1234567-b8c9-0d1e-2fa3-b45c6d7e8912.

$sel:instanceCounts:FleetCapacity', fleetCapacity_instanceCounts - The current instance count and capacity settings for the fleet location.

$sel:location:FleetCapacity', fleetCapacity_location - The fleet location for the instance count information, expressed as an AWS Region code, such as us-west-2.

fleetCapacity_instanceType :: Lens' FleetCapacity (Maybe EC2InstanceType)

The EC2 instance type that is used for all instances in a fleet. The instance type determines the computing resources in use, including CPU, memory, storage, and networking capacity. See Amazon EC2 Instance Types for detailed descriptions.

fleetCapacity_fleetId :: Lens' FleetCapacity (Maybe Text)

A unique identifier for the fleet associated with the location.

fleetCapacity_fleetArn :: Lens' FleetCapacity (Maybe Text)

The Amazon Resource Name (ARN) that is assigned to a GameLift fleet resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:<region>::fleet/fleet-a1234567-b8c9-0d1e-2fa3-b45c6d7e8912.

fleetCapacity_instanceCounts :: Lens' FleetCapacity (Maybe EC2InstanceCounts)

The current instance count and capacity settings for the fleet location.

fleetCapacity_location :: Lens' FleetCapacity (Maybe Text)

The fleet location for the instance count information, expressed as an AWS Region code, such as us-west-2.