Network.AWS.GameLift.Types.FleetUtilization

Description

Documentation

data FleetUtilization

Current resource utilization statistics in a specified fleet or location. The location value might refer to a fleet's remote location or its home Region.

Related actions

DescribeFleetUtilization | DescribeFleetLocationUtilization

See: newFleetUtilization smart constructor.

Constructors

FleetUtilization'

Fields

activeGameSessionCount :: Maybe Natural

The number of active game sessions that are currently being hosted across all instances in the fleet location.

currentPlayerSessionCount :: Maybe Natural

The number of active player sessions that are currently being hosted across all instances in the fleet location.

maximumPlayerSessionCount :: Maybe Natural

The maximum number of players allowed across all game sessions that are currently being hosted across all instances in the fleet location.

fleetId :: Maybe Text

A unique identifier for the fleet associated with the location.

fleetArn :: Maybe Text

The Amazon Resource Name (ARN) that is assigned to a GameLift fleet resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:<region>::fleet/fleet-a1234567-b8c9-0d1e-2fa3-b45c6d7e8912.

activeServerProcessCount :: Maybe Natural

The number of server processes in ACTIVE status that are currently running across all instances in the fleet location.

location :: Maybe Text

The fleet location for the fleet utilization information, expressed as an AWS Region code, such as us-west-2.

newFleetUtilization :: FleetUtilization

Create a value of FleetUtilization with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:activeGameSessionCount:FleetUtilization', fleetUtilization_activeGameSessionCount - The number of active game sessions that are currently being hosted across all instances in the fleet location.

$sel:currentPlayerSessionCount:FleetUtilization', fleetUtilization_currentPlayerSessionCount - The number of active player sessions that are currently being hosted across all instances in the fleet location.

$sel:maximumPlayerSessionCount:FleetUtilization', fleetUtilization_maximumPlayerSessionCount - The maximum number of players allowed across all game sessions that are currently being hosted across all instances in the fleet location.

$sel:fleetId:FleetUtilization', fleetUtilization_fleetId - A unique identifier for the fleet associated with the location.

$sel:fleetArn:FleetUtilization', fleetUtilization_fleetArn - The Amazon Resource Name (ARN) that is assigned to a GameLift fleet resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:<region>::fleet/fleet-a1234567-b8c9-0d1e-2fa3-b45c6d7e8912.

$sel:activeServerProcessCount:FleetUtilization', fleetUtilization_activeServerProcessCount - The number of server processes in ACTIVE status that are currently running across all instances in the fleet location.

$sel:location:FleetUtilization', fleetUtilization_location - The fleet location for the fleet utilization information, expressed as an AWS Region code, such as us-west-2.

fleetUtilization_activeGameSessionCount :: Lens' FleetUtilization (Maybe Natural)

The number of active game sessions that are currently being hosted across all instances in the fleet location.

fleetUtilization_currentPlayerSessionCount :: Lens' FleetUtilization (Maybe Natural)

The number of active player sessions that are currently being hosted across all instances in the fleet location.

fleetUtilization_maximumPlayerSessionCount :: Lens' FleetUtilization (Maybe Natural)

The maximum number of players allowed across all game sessions that are currently being hosted across all instances in the fleet location.

fleetUtilization_fleetId :: Lens' FleetUtilization (Maybe Text)

A unique identifier for the fleet associated with the location.

fleetUtilization_fleetArn :: Lens' FleetUtilization (Maybe Text)

The Amazon Resource Name (ARN) that is assigned to a GameLift fleet resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:<region>::fleet/fleet-a1234567-b8c9-0d1e-2fa3-b45c6d7e8912.

fleetUtilization_activeServerProcessCount :: Lens' FleetUtilization (Maybe Natural)

The number of server processes in ACTIVE status that are currently running across all instances in the fleet location.

fleetUtilization_location :: Lens' FleetUtilization (Maybe Text)

The fleet location for the fleet utilization information, expressed as an AWS Region code, such as us-west-2.