Network.AWS.GameLift.Types.GameServer

Description

Documentation

data GameServer

__This data type is used with the GameLift FleetIQ and game server groups.__

Properties describing a game server that is running on an instance in a GameServerGroup.

A game server is created by a successful call to RegisterGameServer and deleted by calling DeregisterGameServer. A game server is claimed to host a game session by calling ClaimGameServer.

Related actions

RegisterGameServer | ListGameServers | ClaimGameServer | DescribeGameServer | UpdateGameServer | DeregisterGameServer | All APIs by task

See: newGameServer smart constructor.

Constructors

GameServer'

Fields

instanceId :: Maybe Text

The unique identifier for the instance where the game server is running. This ID is available in the instance metadata. EC2 instance IDs use a 17-character format, for example: i-1234567890abcdef0.

utilizationStatus :: Maybe GameServerUtilizationStatus

Indicates whether the game server is currently available for new games or is busy. Possible statuses include:

AVAILABLE - The game server is available to be claimed. A game server that has been claimed remains in this status until it reports game hosting activity.

UTILIZED - The game server is currently hosting a game session with players.

claimStatus :: Maybe GameServerClaimStatus

Indicates when an available game server has been reserved for gameplay but has not yet started hosting a game. Once it is claimed, the game server remains in CLAIMED status for a maximum of one minute. During this time, game clients connect to the game server to start the game and trigger the game server to update its utilization status. After one minute, the game server claim status reverts to null.

gameServerData :: Maybe Text

A set of custom game server properties, formatted as a single string value. This data is passed to a game client or service when it requests information on game servers using ListGameServers or ClaimGameServer.

lastClaimTime :: Maybe POSIX

Timestamp that indicates the last time the game server was claimed with a ClaimGameServer request. The format is a number expressed in Unix time as milliseconds (for example "1469498468.057"). This value is used to calculate when a claimed game server's status should revert to null.

registrationTime :: Maybe POSIX

Timestamp that indicates when the game server was created with a RegisterGameServer request. The format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

gameServerGroupArn :: Maybe Text

The ARN identifier for the game server group where the game server is located.

gameServerId :: Maybe Text

A custom string that uniquely identifies the game server. Game server IDs are developer-defined and are unique across all game server groups in an AWS account.

gameServerGroupName :: Maybe Text

A unique identifier for the game server group where the game server is running. Use either the GameServerGroup name or ARN value.

connectionInfo :: Maybe Text

The port and IP address that must be used to establish a client connection to the game server.

lastHealthCheckTime :: Maybe POSIX

Timestamp that indicates the last time the game server was updated with health status using an UpdateGameServer request. The format is a number expressed in Unix time as milliseconds (for example "1469498468.057"). After game server registration, this property is only changed when a game server update specifies a health check value.

newGameServer :: GameServer

Create a value of GameServer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceId:GameServer', gameServer_instanceId - The unique identifier for the instance where the game server is running. This ID is available in the instance metadata. EC2 instance IDs use a 17-character format, for example: i-1234567890abcdef0.

$sel:utilizationStatus:GameServer', gameServer_utilizationStatus - Indicates whether the game server is currently available for new games or is busy. Possible statuses include:

AVAILABLE - The game server is available to be claimed. A game server that has been claimed remains in this status until it reports game hosting activity.

UTILIZED - The game server is currently hosting a game session with players.

$sel:claimStatus:GameServer', gameServer_claimStatus - Indicates when an available game server has been reserved for gameplay but has not yet started hosting a game. Once it is claimed, the game server remains in CLAIMED status for a maximum of one minute. During this time, game clients connect to the game server to start the game and trigger the game server to update its utilization status. After one minute, the game server claim status reverts to null.

$sel:gameServerData:GameServer', gameServer_gameServerData - A set of custom game server properties, formatted as a single string value. This data is passed to a game client or service when it requests information on game servers using ListGameServers or ClaimGameServer.

$sel:lastClaimTime:GameServer', gameServer_lastClaimTime - Timestamp that indicates the last time the game server was claimed with a ClaimGameServer request. The format is a number expressed in Unix time as milliseconds (for example "1469498468.057"). This value is used to calculate when a claimed game server's status should revert to null.

$sel:registrationTime:GameServer', gameServer_registrationTime - Timestamp that indicates when the game server was created with a RegisterGameServer request. The format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

$sel:gameServerGroupArn:GameServer', gameServer_gameServerGroupArn - The ARN identifier for the game server group where the game server is located.

$sel:gameServerId:GameServer', gameServer_gameServerId - A custom string that uniquely identifies the game server. Game server IDs are developer-defined and are unique across all game server groups in an AWS account.

$sel:gameServerGroupName:GameServer', gameServer_gameServerGroupName - A unique identifier for the game server group where the game server is running. Use either the GameServerGroup name or ARN value.

$sel:connectionInfo:GameServer', gameServer_connectionInfo - The port and IP address that must be used to establish a client connection to the game server.

$sel:lastHealthCheckTime:GameServer', gameServer_lastHealthCheckTime - Timestamp that indicates the last time the game server was updated with health status using an UpdateGameServer request. The format is a number expressed in Unix time as milliseconds (for example "1469498468.057"). After game server registration, this property is only changed when a game server update specifies a health check value.

gameServer_instanceId :: Lens' GameServer (Maybe Text)

The unique identifier for the instance where the game server is running. This ID is available in the instance metadata. EC2 instance IDs use a 17-character format, for example: i-1234567890abcdef0.

gameServer_utilizationStatus :: Lens' GameServer (Maybe GameServerUtilizationStatus)

Indicates whether the game server is currently available for new games or is busy. Possible statuses include:

AVAILABLE - The game server is available to be claimed. A game server that has been claimed remains in this status until it reports game hosting activity.

UTILIZED - The game server is currently hosting a game session with players.

gameServer_claimStatus :: Lens' GameServer (Maybe GameServerClaimStatus)

Indicates when an available game server has been reserved for gameplay but has not yet started hosting a game. Once it is claimed, the game server remains in CLAIMED status for a maximum of one minute. During this time, game clients connect to the game server to start the game and trigger the game server to update its utilization status. After one minute, the game server claim status reverts to null.

gameServer_gameServerData :: Lens' GameServer (Maybe Text)

A set of custom game server properties, formatted as a single string value. This data is passed to a game client or service when it requests information on game servers using ListGameServers or ClaimGameServer.

gameServer_lastClaimTime :: Lens' GameServer (Maybe UTCTime)

Timestamp that indicates the last time the game server was claimed with a ClaimGameServer request. The format is a number expressed in Unix time as milliseconds (for example "1469498468.057"). This value is used to calculate when a claimed game server's status should revert to null.

gameServer_registrationTime :: Lens' GameServer (Maybe UTCTime)

Timestamp that indicates when the game server was created with a RegisterGameServer request. The format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

gameServer_gameServerGroupArn :: Lens' GameServer (Maybe Text)

The ARN identifier for the game server group where the game server is located.

gameServer_gameServerId :: Lens' GameServer (Maybe Text)

A custom string that uniquely identifies the game server. Game server IDs are developer-defined and are unique across all game server groups in an AWS account.

gameServer_gameServerGroupName :: Lens' GameServer (Maybe Text)

A unique identifier for the game server group where the game server is running. Use either the GameServerGroup name or ARN value.

gameServer_connectionInfo :: Lens' GameServer (Maybe Text)

The port and IP address that must be used to establish a client connection to the game server.

gameServer_lastHealthCheckTime :: Lens' GameServer (Maybe UTCTime)

Timestamp that indicates the last time the game server was updated with health status using an UpdateGameServer request. The format is a number expressed in Unix time as milliseconds (for example "1469498468.057"). After game server registration, this property is only changed when a game server update specifies a health check value.