Network.AWS.GameLift.Types.GameServerGroup

Description

Documentation

data GameServerGroup

__This data type is used with the GameLift FleetIQ and game server groups.__

Properties that describe a game server group resource. A game server group manages certain properties related to a corresponding EC2 Auto Scaling group.

A game server group is created by a successful call to CreateGameServerGroup and deleted by calling DeleteGameServerGroup. Game server group activity can be temporarily suspended and resumed by calling SuspendGameServerGroup and ResumeGameServerGroup, respectively.

Related actions

CreateGameServerGroup | ListGameServerGroups | DescribeGameServerGroup | UpdateGameServerGroup | DeleteGameServerGroup | ResumeGameServerGroup | SuspendGameServerGroup | DescribeGameServerInstances | All APIs by task

See: newGameServerGroup smart constructor.

Constructors

GameServerGroup'

Fields

creationTime :: Maybe POSIX

A time stamp indicating when this data object was created. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

status :: Maybe GameServerGroupStatus

The current status of the game server group. Possible statuses include:

NEW - GameLift FleetIQ has validated the CreateGameServerGroup() request.

ACTIVATING - GameLift FleetIQ is setting up a game server group, which includes creating an Auto Scaling group in your AWS account.

ACTIVE - The game server group has been successfully created.

DELETE_SCHEDULED - A request to delete the game server group has been received.

DELETING - GameLift FleetIQ has received a valid DeleteGameServerGroup() request and is processing it. GameLift FleetIQ must first complete and release hosts before it deletes the Auto Scaling group and the game server group.

DELETED - The game server group has been successfully deleted.

ERROR - The asynchronous processes of activating or deleting a game server group has failed, resulting in an error state.

roleArn :: Maybe Text

The Amazon Resource Name (ARN) for an IAM role that allows Amazon GameLift to access your EC2 Auto Scaling groups.

autoScalingGroupArn :: Maybe Text

A generated unique ID for the EC2 Auto Scaling group that is associated with this game server group.

instanceDefinitions :: Maybe (NonEmpty InstanceDefinition)

The set of EC2 instance types that GameLift FleetIQ can use when balancing and automatically scaling instances in the corresponding Auto Scaling group.

gameServerGroupArn :: Maybe Text

A generated unique ID for the game server group.

suspendedActions :: Maybe (NonEmpty GameServerGroupAction)

A list of activities that are currently suspended for this game server group. If this property is empty, all activities are occurring.

gameServerGroupName :: Maybe Text

A developer-defined identifier for the game server group. The name is unique for each Region in each AWS account.

balancingStrategy :: Maybe BalancingStrategy

Indicates how GameLift FleetIQ balances the use of Spot Instances and On-Demand Instances in the game server group. Method options include the following:

SPOT_ONLY - Only Spot Instances are used in the game server group. If Spot Instances are unavailable or not viable for game hosting, the game server group provides no hosting capacity until Spot Instances can again be used. Until then, no new instances are started, and the existing nonviable Spot Instances are terminated (after current gameplay ends) and are not replaced.

SPOT_PREFERRED - (default value) Spot Instances are used whenever available in the game server group. If Spot Instances are unavailable, the game server group continues to provide hosting capacity by falling back to On-Demand Instances. Existing nonviable Spot Instances are terminated (after current gameplay ends) and are replaced with new On-Demand Instances.

ON_DEMAND_ONLY - Only On-Demand Instances are used in the game server group. No Spot Instances are used, even when available, while this balancing strategy is in force.

statusReason :: Maybe Text

Additional information about the current game server group status. This information might provide additional insight on groups that are in ERROR status.

gameServerProtectionPolicy :: Maybe GameServerProtectionPolicy

A flag that indicates whether instances in the game server group are protected from early termination. Unprotected instances that have active game servers running might be terminated during a scale-down event, causing players to be dropped from the game. Protected instances cannot be terminated while there are active game servers running except in the event of a forced game server group deletion (see ). An exception to this is with Spot Instances, which can be terminated by AWS regardless of protection status.

lastUpdatedTime :: Maybe POSIX

A timestamp that indicates when this game server group was last updated.

newGameServerGroup :: GameServerGroup

Create a value of GameServerGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:GameServerGroup', gameServerGroup_creationTime - A time stamp indicating when this data object was created. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

$sel:status:GameServerGroup', gameServerGroup_status - The current status of the game server group. Possible statuses include:

NEW - GameLift FleetIQ has validated the CreateGameServerGroup() request.

ACTIVATING - GameLift FleetIQ is setting up a game server group, which includes creating an Auto Scaling group in your AWS account.

ACTIVE - The game server group has been successfully created.

DELETE_SCHEDULED - A request to delete the game server group has been received.

DELETING - GameLift FleetIQ has received a valid DeleteGameServerGroup() request and is processing it. GameLift FleetIQ must first complete and release hosts before it deletes the Auto Scaling group and the game server group.

DELETED - The game server group has been successfully deleted.

ERROR - The asynchronous processes of activating or deleting a game server group has failed, resulting in an error state.

$sel:roleArn:GameServerGroup', gameServerGroup_roleArn - The Amazon Resource Name (ARN) for an IAM role that allows Amazon GameLift to access your EC2 Auto Scaling groups.

$sel:autoScalingGroupArn:GameServerGroup', gameServerGroup_autoScalingGroupArn - A generated unique ID for the EC2 Auto Scaling group that is associated with this game server group.

$sel:instanceDefinitions:GameServerGroup', gameServerGroup_instanceDefinitions - The set of EC2 instance types that GameLift FleetIQ can use when balancing and automatically scaling instances in the corresponding Auto Scaling group.

$sel:gameServerGroupArn:GameServerGroup', gameServerGroup_gameServerGroupArn - A generated unique ID for the game server group.

$sel:suspendedActions:GameServerGroup', gameServerGroup_suspendedActions - A list of activities that are currently suspended for this game server group. If this property is empty, all activities are occurring.

$sel:gameServerGroupName:GameServerGroup', gameServerGroup_gameServerGroupName - A developer-defined identifier for the game server group. The name is unique for each Region in each AWS account.

$sel:balancingStrategy:GameServerGroup', gameServerGroup_balancingStrategy - Indicates how GameLift FleetIQ balances the use of Spot Instances and On-Demand Instances in the game server group. Method options include the following:

SPOT_ONLY - Only Spot Instances are used in the game server group. If Spot Instances are unavailable or not viable for game hosting, the game server group provides no hosting capacity until Spot Instances can again be used. Until then, no new instances are started, and the existing nonviable Spot Instances are terminated (after current gameplay ends) and are not replaced.

SPOT_PREFERRED - (default value) Spot Instances are used whenever available in the game server group. If Spot Instances are unavailable, the game server group continues to provide hosting capacity by falling back to On-Demand Instances. Existing nonviable Spot Instances are terminated (after current gameplay ends) and are replaced with new On-Demand Instances.

ON_DEMAND_ONLY - Only On-Demand Instances are used in the game server group. No Spot Instances are used, even when available, while this balancing strategy is in force.

$sel:statusReason:GameServerGroup', gameServerGroup_statusReason - Additional information about the current game server group status. This information might provide additional insight on groups that are in ERROR status.

$sel:gameServerProtectionPolicy:GameServerGroup', gameServerGroup_gameServerProtectionPolicy - A flag that indicates whether instances in the game server group are protected from early termination. Unprotected instances that have active game servers running might be terminated during a scale-down event, causing players to be dropped from the game. Protected instances cannot be terminated while there are active game servers running except in the event of a forced game server group deletion (see ). An exception to this is with Spot Instances, which can be terminated by AWS regardless of protection status.

$sel:lastUpdatedTime:GameServerGroup', gameServerGroup_lastUpdatedTime - A timestamp that indicates when this game server group was last updated.

gameServerGroup_creationTime :: Lens' GameServerGroup (Maybe UTCTime)

A time stamp indicating when this data object was created. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

gameServerGroup_status :: Lens' GameServerGroup (Maybe GameServerGroupStatus)

The current status of the game server group. Possible statuses include:

NEW - GameLift FleetIQ has validated the CreateGameServerGroup() request.

ACTIVATING - GameLift FleetIQ is setting up a game server group, which includes creating an Auto Scaling group in your AWS account.

ACTIVE - The game server group has been successfully created.

DELETE_SCHEDULED - A request to delete the game server group has been received.

DELETING - GameLift FleetIQ has received a valid DeleteGameServerGroup() request and is processing it. GameLift FleetIQ must first complete and release hosts before it deletes the Auto Scaling group and the game server group.

DELETED - The game server group has been successfully deleted.

ERROR - The asynchronous processes of activating or deleting a game server group has failed, resulting in an error state.

gameServerGroup_roleArn :: Lens' GameServerGroup (Maybe Text)

The Amazon Resource Name (ARN) for an IAM role that allows Amazon GameLift to access your EC2 Auto Scaling groups.

gameServerGroup_autoScalingGroupArn :: Lens' GameServerGroup (Maybe Text)

A generated unique ID for the EC2 Auto Scaling group that is associated with this game server group.

gameServerGroup_instanceDefinitions :: Lens' GameServerGroup (Maybe (NonEmpty InstanceDefinition))

The set of EC2 instance types that GameLift FleetIQ can use when balancing and automatically scaling instances in the corresponding Auto Scaling group.

gameServerGroup_gameServerGroupArn :: Lens' GameServerGroup (Maybe Text)

A generated unique ID for the game server group.

gameServerGroup_suspendedActions :: Lens' GameServerGroup (Maybe (NonEmpty GameServerGroupAction))

A list of activities that are currently suspended for this game server group. If this property is empty, all activities are occurring.

gameServerGroup_gameServerGroupName :: Lens' GameServerGroup (Maybe Text)

A developer-defined identifier for the game server group. The name is unique for each Region in each AWS account.

gameServerGroup_balancingStrategy :: Lens' GameServerGroup (Maybe BalancingStrategy)

Indicates how GameLift FleetIQ balances the use of Spot Instances and On-Demand Instances in the game server group. Method options include the following:

SPOT_ONLY - Only Spot Instances are used in the game server group. If Spot Instances are unavailable or not viable for game hosting, the game server group provides no hosting capacity until Spot Instances can again be used. Until then, no new instances are started, and the existing nonviable Spot Instances are terminated (after current gameplay ends) and are not replaced.

SPOT_PREFERRED - (default value) Spot Instances are used whenever available in the game server group. If Spot Instances are unavailable, the game server group continues to provide hosting capacity by falling back to On-Demand Instances. Existing nonviable Spot Instances are terminated (after current gameplay ends) and are replaced with new On-Demand Instances.

ON_DEMAND_ONLY - Only On-Demand Instances are used in the game server group. No Spot Instances are used, even when available, while this balancing strategy is in force.

gameServerGroup_statusReason :: Lens' GameServerGroup (Maybe Text)

Additional information about the current game server group status. This information might provide additional insight on groups that are in ERROR status.

gameServerGroup_gameServerProtectionPolicy :: Lens' GameServerGroup (Maybe GameServerProtectionPolicy)

A flag that indicates whether instances in the game server group are protected from early termination. Unprotected instances that have active game servers running might be terminated during a scale-down event, causing players to be dropped from the game. Protected instances cannot be terminated while there are active game servers running except in the event of a forced game server group deletion (see ). An exception to this is with Spot Instances, which can be terminated by AWS regardless of protection status.

gameServerGroup_lastUpdatedTime :: Lens' GameServerGroup (Maybe UTCTime)

A timestamp that indicates when this game server group was last updated.