Network.AWS.GameLift.Types.GameServerInstance

Description

Documentation

data GameServerInstance

__This data type is used with the GameLift FleetIQ and game server groups.__

Additional properties, including status, that describe an EC2 instance in a game server group. Instance configurations are set with game server group properties (see DescribeGameServerGroup and with the EC2 launch template that was used when creating the game server group.

Retrieve game server instances for a game server group by calling DescribeGameServerInstances.

Related actions

CreateGameServerGroup | ListGameServerGroups | DescribeGameServerGroup | UpdateGameServerGroup | DeleteGameServerGroup | ResumeGameServerGroup | SuspendGameServerGroup | DescribeGameServerInstances | All APIs by task

See: newGameServerInstance smart constructor.

Constructors

GameServerInstance'

Fields

instanceId :: Maybe Text

The unique identifier for the instance where the game server is running. This ID is available in the instance metadata. EC2 instance IDs use a 17-character format, for example: i-1234567890abcdef0.

instanceStatus :: Maybe GameServerInstanceStatus

Current status of the game server instance.

ACTIVE -- The instance is viable for hosting game servers.

DRAINING -- The instance is not viable for hosting game servers. Existing game servers are in the process of ending, and new game servers are not started on this instance unless no other resources are available. When the instance is put in DRAINING, a new instance is started up to replace it. Once the instance has no UTILIZED game servers, it will be terminated in favor of the new instance.

SPOT_TERMINATING -- The instance is in the process of shutting down due to a Spot instance interruption. No new game servers are started on this instance.

gameServerGroupArn :: Maybe Text

A generated unique identifier for the game server group that includes the game server instance.

gameServerGroupName :: Maybe Text

A developer-defined identifier for the game server group that includes the game server instance. The name is unique for each Region in each AWS account.

newGameServerInstance :: GameServerInstance

Create a value of GameServerInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceId:GameServerInstance', gameServerInstance_instanceId - The unique identifier for the instance where the game server is running. This ID is available in the instance metadata. EC2 instance IDs use a 17-character format, for example: i-1234567890abcdef0.

$sel:instanceStatus:GameServerInstance', gameServerInstance_instanceStatus - Current status of the game server instance.

ACTIVE -- The instance is viable for hosting game servers.

DRAINING -- The instance is not viable for hosting game servers. Existing game servers are in the process of ending, and new game servers are not started on this instance unless no other resources are available. When the instance is put in DRAINING, a new instance is started up to replace it. Once the instance has no UTILIZED game servers, it will be terminated in favor of the new instance.

SPOT_TERMINATING -- The instance is in the process of shutting down due to a Spot instance interruption. No new game servers are started on this instance.

$sel:gameServerGroupArn:GameServerInstance', gameServerInstance_gameServerGroupArn - A generated unique identifier for the game server group that includes the game server instance.

$sel:gameServerGroupName:GameServerInstance', gameServerInstance_gameServerGroupName - A developer-defined identifier for the game server group that includes the game server instance. The name is unique for each Region in each AWS account.

gameServerInstance_instanceId :: Lens' GameServerInstance (Maybe Text)

The unique identifier for the instance where the game server is running. This ID is available in the instance metadata. EC2 instance IDs use a 17-character format, for example: i-1234567890abcdef0.

gameServerInstance_instanceStatus :: Lens' GameServerInstance (Maybe GameServerInstanceStatus)

Current status of the game server instance.

ACTIVE -- The instance is viable for hosting game servers.

DRAINING -- The instance is not viable for hosting game servers. Existing game servers are in the process of ending, and new game servers are not started on this instance unless no other resources are available. When the instance is put in DRAINING, a new instance is started up to replace it. Once the instance has no UTILIZED game servers, it will be terminated in favor of the new instance.

SPOT_TERMINATING -- The instance is in the process of shutting down due to a Spot instance interruption. No new game servers are started on this instance.

gameServerInstance_gameServerGroupArn :: Lens' GameServerInstance (Maybe Text)

A generated unique identifier for the game server group that includes the game server instance.

gameServerInstance_gameServerGroupName :: Lens' GameServerInstance (Maybe Text)

A developer-defined identifier for the game server group that includes the game server instance. The name is unique for each Region in each AWS account.