Network.AWS.GameLift.Types.GameSession

Description

Documentation

data GameSession

Properties describing a game session.

A game session in ACTIVE status can host players. When a game session ends, its status is set to TERMINATED.

Once the session ends, the game session object is retained for 30 days. This means you can reuse idempotency token values after this time. Game session logs are retained for 14 days.

Related actions

CreateGameSession | DescribeGameSessions | DescribeGameSessionDetails | SearchGameSessions | UpdateGameSession | GetGameSessionLogUrl | StartGameSessionPlacement | DescribeGameSessionPlacement | StopGameSessionPlacement | All APIs by task

See: newGameSession smart constructor.

Constructors

GameSession'

Fields

currentPlayerSessionCount :: Maybe Natural

Number of players currently in the game session.

gameProperties :: Maybe [GameProperty]

A set of custom properties for a game session, formatted as key:value pairs. These properties are passed to a game server process in the GameSession object with a request to start a new game session. You can search for active game sessions based on this custom data with SearchGameSessions.

creationTime :: Maybe POSIX

A time stamp indicating when this data object was created. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

status :: Maybe GameSessionStatus

Current status of the game session. A game session must have an ACTIVE status to have player sessions.

playerSessionCreationPolicy :: Maybe PlayerSessionCreationPolicy

Indicates whether or not the game session is accepting new players.

creatorId :: Maybe Text

A unique identifier for a player. This ID is used to enforce a resource protection policy (if one exists), that limits the number of game sessions a player can create.

maximumPlayerSessionCount :: Maybe Natural

The maximum number of players that can be connected simultaneously to the game session.

matchmakerData :: Maybe Text

Information about the matchmaking process that was used to create the game session. It is in JSON syntax, formatted as a string. In addition the matchmaking configuration used, it contains data on all players assigned to the match, including player attributes and team assignments. For more details on matchmaker data, see Match Data. Matchmaker data is useful when requesting match backfills, and is updated whenever new players are added during a successful backfill (see StartMatchBackfill).

fleetId :: Maybe Text

A unique identifier for the fleet that the game session is running on.

fleetArn :: Maybe Text

The Amazon Resource Name (ARN) associated with the GameLift fleet that this game session is running on.

gameSessionData :: Maybe Text

A set of custom game session properties, formatted as a single string value. This data is passed to a game server process in the GameSession object with a request to start a new game session.

gameSessionId :: Maybe Text

A unique identifier for the game session. A game session ARN has the following format: arn:aws:gamelift:<region>::gamesession/<fleet ID>/<custom ID string or idempotency token>.

name :: Maybe Text

A descriptive label that is associated with a game session. Session names do not need to be unique.

ipAddress :: Maybe Text

The IP address of the game session. To connect to a GameLift game server, an app needs both the IP address and port number.

terminationTime :: Maybe POSIX

A time stamp indicating when this data object was terminated. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

port :: Maybe Natural

The port number for the game session. To connect to a GameLift game server, an app needs both the IP address and port number.

dnsName :: Maybe Text

The DNS identifier assigned to the instance that is running the game session. Values have the following format:

TLS-enabled fleets: <unique identifier>.<region identifier>.amazongamelift.com.

Non-TLS-enabled fleets: ec2-<unique identifier>.compute.amazonaws.com. (See Amazon EC2 Instance IP Addressing.)

When connecting to a game session that is running on a TLS-enabled fleet, you must use the DNS name, not the IP address.

statusReason :: Maybe GameSessionStatusReason

Provides additional information about game session status. INTERRUPTED indicates that the game session was hosted on a spot instance that was reclaimed, causing the active game session to be terminated.

location :: Maybe Text

The fleet location where the game session is running. This value might specify the fleet's home Region or a remote location. Location is expressed as an AWS Region code such as us-west-2.

newGameSession :: GameSession

Create a value of GameSession with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:currentPlayerSessionCount:GameSession', gameSession_currentPlayerSessionCount - Number of players currently in the game session.

$sel:gameProperties:GameSession', gameSession_gameProperties - A set of custom properties for a game session, formatted as key:value pairs. These properties are passed to a game server process in the GameSession object with a request to start a new game session. You can search for active game sessions based on this custom data with SearchGameSessions.

$sel:creationTime:GameSession', gameSession_creationTime - A time stamp indicating when this data object was created. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

$sel:status:GameSession', gameSession_status - Current status of the game session. A game session must have an ACTIVE status to have player sessions.

$sel:playerSessionCreationPolicy:GameSession', gameSession_playerSessionCreationPolicy - Indicates whether or not the game session is accepting new players.

$sel:creatorId:GameSession', gameSession_creatorId - A unique identifier for a player. This ID is used to enforce a resource protection policy (if one exists), that limits the number of game sessions a player can create.

$sel:maximumPlayerSessionCount:GameSession', gameSession_maximumPlayerSessionCount - The maximum number of players that can be connected simultaneously to the game session.

$sel:matchmakerData:GameSession', gameSession_matchmakerData - Information about the matchmaking process that was used to create the game session. It is in JSON syntax, formatted as a string. In addition the matchmaking configuration used, it contains data on all players assigned to the match, including player attributes and team assignments. For more details on matchmaker data, see Match Data. Matchmaker data is useful when requesting match backfills, and is updated whenever new players are added during a successful backfill (see StartMatchBackfill).

$sel:fleetId:GameSession', gameSession_fleetId - A unique identifier for the fleet that the game session is running on.

$sel:fleetArn:GameSession', gameSession_fleetArn - The Amazon Resource Name (ARN) associated with the GameLift fleet that this game session is running on.

$sel:gameSessionData:GameSession', gameSession_gameSessionData - A set of custom game session properties, formatted as a single string value. This data is passed to a game server process in the GameSession object with a request to start a new game session.

$sel:gameSessionId:GameSession', gameSession_gameSessionId - A unique identifier for the game session. A game session ARN has the following format: arn:aws:gamelift:<region>::gamesession/<fleet ID>/<custom ID string or idempotency token>.

$sel:name:GameSession', gameSession_name - A descriptive label that is associated with a game session. Session names do not need to be unique.

$sel:ipAddress:GameSession', gameSession_ipAddress - The IP address of the game session. To connect to a GameLift game server, an app needs both the IP address and port number.

$sel:terminationTime:GameSession', gameSession_terminationTime - A time stamp indicating when this data object was terminated. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

$sel:port:GameSession', gameSession_port - The port number for the game session. To connect to a GameLift game server, an app needs both the IP address and port number.

$sel:dnsName:GameSession', gameSession_dnsName - The DNS identifier assigned to the instance that is running the game session. Values have the following format:

TLS-enabled fleets: <unique identifier>.<region identifier>.amazongamelift.com.

Non-TLS-enabled fleets: ec2-<unique identifier>.compute.amazonaws.com. (See Amazon EC2 Instance IP Addressing.)

When connecting to a game session that is running on a TLS-enabled fleet, you must use the DNS name, not the IP address.

$sel:statusReason:GameSession', gameSession_statusReason - Provides additional information about game session status. INTERRUPTED indicates that the game session was hosted on a spot instance that was reclaimed, causing the active game session to be terminated.

$sel:location:GameSession', gameSession_location - The fleet location where the game session is running. This value might specify the fleet's home Region or a remote location. Location is expressed as an AWS Region code such as us-west-2.

gameSession_currentPlayerSessionCount :: Lens' GameSession (Maybe Natural)

Number of players currently in the game session.

gameSession_gameProperties :: Lens' GameSession (Maybe [GameProperty])

A set of custom properties for a game session, formatted as key:value pairs. These properties are passed to a game server process in the GameSession object with a request to start a new game session. You can search for active game sessions based on this custom data with SearchGameSessions.

gameSession_creationTime :: Lens' GameSession (Maybe UTCTime)

A time stamp indicating when this data object was created. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

gameSession_status :: Lens' GameSession (Maybe GameSessionStatus)

Current status of the game session. A game session must have an ACTIVE status to have player sessions.

gameSession_playerSessionCreationPolicy :: Lens' GameSession (Maybe PlayerSessionCreationPolicy)

Indicates whether or not the game session is accepting new players.

gameSession_creatorId :: Lens' GameSession (Maybe Text)

A unique identifier for a player. This ID is used to enforce a resource protection policy (if one exists), that limits the number of game sessions a player can create.

gameSession_maximumPlayerSessionCount :: Lens' GameSession (Maybe Natural)

The maximum number of players that can be connected simultaneously to the game session.

gameSession_matchmakerData :: Lens' GameSession (Maybe Text)

Information about the matchmaking process that was used to create the game session. It is in JSON syntax, formatted as a string. In addition the matchmaking configuration used, it contains data on all players assigned to the match, including player attributes and team assignments. For more details on matchmaker data, see Match Data. Matchmaker data is useful when requesting match backfills, and is updated whenever new players are added during a successful backfill (see StartMatchBackfill).

gameSession_fleetId :: Lens' GameSession (Maybe Text)

A unique identifier for the fleet that the game session is running on.

gameSession_fleetArn :: Lens' GameSession (Maybe Text)

The Amazon Resource Name (ARN) associated with the GameLift fleet that this game session is running on.

gameSession_gameSessionData :: Lens' GameSession (Maybe Text)

A set of custom game session properties, formatted as a single string value. This data is passed to a game server process in the GameSession object with a request to start a new game session.

gameSession_gameSessionId :: Lens' GameSession (Maybe Text)

A unique identifier for the game session. A game session ARN has the following format: arn:aws:gamelift:<region>::gamesession/<fleet ID>/<custom ID string or idempotency token>.

gameSession_name :: Lens' GameSession (Maybe Text)

A descriptive label that is associated with a game session. Session names do not need to be unique.

gameSession_ipAddress :: Lens' GameSession (Maybe Text)

The IP address of the game session. To connect to a GameLift game server, an app needs both the IP address and port number.

gameSession_terminationTime :: Lens' GameSession (Maybe UTCTime)

A time stamp indicating when this data object was terminated. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

gameSession_port :: Lens' GameSession (Maybe Natural)

The port number for the game session. To connect to a GameLift game server, an app needs both the IP address and port number.

gameSession_dnsName :: Lens' GameSession (Maybe Text)

The DNS identifier assigned to the instance that is running the game session. Values have the following format:

TLS-enabled fleets: <unique identifier>.<region identifier>.amazongamelift.com.

Non-TLS-enabled fleets: ec2-<unique identifier>.compute.amazonaws.com. (See Amazon EC2 Instance IP Addressing.)

When connecting to a game session that is running on a TLS-enabled fleet, you must use the DNS name, not the IP address.

gameSession_statusReason :: Lens' GameSession (Maybe GameSessionStatusReason)

Provides additional information about game session status. INTERRUPTED indicates that the game session was hosted on a spot instance that was reclaimed, causing the active game session to be terminated.

gameSession_location :: Lens' GameSession (Maybe Text)

The fleet location where the game session is running. This value might specify the fleet's home Region or a remote location. Location is expressed as an AWS Region code such as us-west-2.