Network.AWS.GameLift.Types.GameSessionPlacement

Description

Documentation

data GameSessionPlacement

Object that describes a StartGameSessionPlacement request. This object includes the full details of the original request plus the current status and start/end time stamps.

Game session placement-related operations include:

StartGameSessionPlacement

DescribeGameSessionPlacement

StopGameSessionPlacement

See: newGameSessionPlacement smart constructor.

Constructors

GameSessionPlacement'

Fields

gameProperties :: Maybe [GameProperty]

A set of custom properties for a game session, formatted as key:value pairs. These properties are passed to a game server process in the GameSession object with a request to start a new game session (see Start a Game Session).

status :: Maybe GameSessionPlacementState

Current status of the game session placement request.

PENDING -- The placement request is currently in the queue waiting to be processed.

FULFILLED -- A new game session and player sessions (if requested) have been successfully created. Values for GameSessionArn and GameSessionRegion are available.

CANCELLED -- The placement request was canceled with a call to StopGameSessionPlacement.

TIMED_OUT -- A new game session was not successfully created before the time limit expired. You can resubmit the placement request as needed.

FAILED -- GameLift is not able to complete the process of placing the game session. Common reasons are the game session terminated before the placement process was completed, or an unexpected internal error.

gameSessionQueueName :: Maybe Text

A descriptive label that is associated with game session queue. Queue names must be unique within each Region.

maximumPlayerSessionCount :: Maybe Natural

The maximum number of players that can be connected simultaneously to the game session.

matchmakerData :: Maybe Text

Information on the matchmaking process for this game. Data is in JSON syntax, formatted as a string. It identifies the matchmaking configuration used to create the match, and contains data on all players assigned to the match, including player attributes and team assignments. For more details on matchmaker data, see Match Data.

gameSessionData :: Maybe Text

A set of custom game session properties, formatted as a single string value. This data is passed to a game server process in the GameSession object with a request to start a new game session (see Start a Game Session).

gameSessionId :: Maybe Text

A unique identifier for the game session. This value is set once the new game session is placed (placement status is FULFILLED).

startTime :: Maybe POSIX

Time stamp indicating when this request was placed in the queue. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

gameSessionArn :: Maybe Text

Identifier for the game session created by this placement request. This value is set once the new game session is placed (placement status is FULFILLED). This identifier is unique across all Regions. You can use this value as a GameSessionId value as needed.

endTime :: Maybe POSIX

Time stamp indicating when this request was completed, canceled, or timed out.

gameSessionName :: Maybe Text

A descriptive label that is associated with a game session. Session names do not need to be unique.

ipAddress :: Maybe Text

The IP address of the game session. To connect to a GameLift game server, an app needs both the IP address and port number. This value is set once the new game session is placed (placement status is FULFILLED).

placementId :: Maybe Text

A unique identifier for a game session placement.

placedPlayerSessions :: Maybe [PlacedPlayerSession]

A collection of information on player sessions created in response to the game session placement request. These player sessions are created only once a new game session is successfully placed (placement status is FULFILLED). This information includes the player ID (as provided in the placement request) and the corresponding player session ID. Retrieve full player sessions by calling DescribePlayerSessions with the player session ID.

port :: Maybe Natural

The port number for the game session. To connect to a GameLift game server, an app needs both the IP address and port number. This value is set once the new game session is placed (placement status is FULFILLED).

dnsName :: Maybe Text

The DNS identifier assigned to the instance that is running the game session. Values have the following format:

TLS-enabled fleets: <unique identifier>.<region identifier>.amazongamelift.com.

Non-TLS-enabled fleets: ec2-<unique identifier>.compute.amazonaws.com. (See Amazon EC2 Instance IP Addressing.)

When connecting to a game session that is running on a TLS-enabled fleet, you must use the DNS name, not the IP address.

gameSessionRegion :: Maybe Text

Name of the Region where the game session created by this placement request is running. This value is set once the new game session is placed (placement status is FULFILLED).

playerLatencies :: Maybe [PlayerLatency]

A set of values, expressed in milliseconds, that indicates the amount of latency that a player experiences when connected to AWS Regions.

newGameSessionPlacement :: GameSessionPlacement

Create a value of GameSessionPlacement with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:gameProperties:GameSessionPlacement', gameSessionPlacement_gameProperties - A set of custom properties for a game session, formatted as key:value pairs. These properties are passed to a game server process in the GameSession object with a request to start a new game session (see Start a Game Session).

$sel:status:GameSessionPlacement', gameSessionPlacement_status - Current status of the game session placement request.

PENDING -- The placement request is currently in the queue waiting to be processed.

FULFILLED -- A new game session and player sessions (if requested) have been successfully created. Values for GameSessionArn and GameSessionRegion are available.

CANCELLED -- The placement request was canceled with a call to StopGameSessionPlacement.

TIMED_OUT -- A new game session was not successfully created before the time limit expired. You can resubmit the placement request as needed.

FAILED -- GameLift is not able to complete the process of placing the game session. Common reasons are the game session terminated before the placement process was completed, or an unexpected internal error.

$sel:gameSessionQueueName:GameSessionPlacement', gameSessionPlacement_gameSessionQueueName - A descriptive label that is associated with game session queue. Queue names must be unique within each Region.

$sel:maximumPlayerSessionCount:GameSessionPlacement', gameSessionPlacement_maximumPlayerSessionCount - The maximum number of players that can be connected simultaneously to the game session.

$sel:matchmakerData:GameSessionPlacement', gameSessionPlacement_matchmakerData - Information on the matchmaking process for this game. Data is in JSON syntax, formatted as a string. It identifies the matchmaking configuration used to create the match, and contains data on all players assigned to the match, including player attributes and team assignments. For more details on matchmaker data, see Match Data.

$sel:gameSessionData:GameSessionPlacement', gameSessionPlacement_gameSessionData - A set of custom game session properties, formatted as a single string value. This data is passed to a game server process in the GameSession object with a request to start a new game session (see Start a Game Session).

$sel:gameSessionId:GameSessionPlacement', gameSessionPlacement_gameSessionId - A unique identifier for the game session. This value is set once the new game session is placed (placement status is FULFILLED).

$sel:startTime:GameSessionPlacement', gameSessionPlacement_startTime - Time stamp indicating when this request was placed in the queue. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

$sel:gameSessionArn:GameSessionPlacement', gameSessionPlacement_gameSessionArn - Identifier for the game session created by this placement request. This value is set once the new game session is placed (placement status is FULFILLED). This identifier is unique across all Regions. You can use this value as a GameSessionId value as needed.

$sel:endTime:GameSessionPlacement', gameSessionPlacement_endTime - Time stamp indicating when this request was completed, canceled, or timed out.

$sel:gameSessionName:GameSessionPlacement', gameSessionPlacement_gameSessionName - A descriptive label that is associated with a game session. Session names do not need to be unique.

$sel:ipAddress:GameSessionPlacement', gameSessionPlacement_ipAddress - The IP address of the game session. To connect to a GameLift game server, an app needs both the IP address and port number. This value is set once the new game session is placed (placement status is FULFILLED).

$sel:placementId:GameSessionPlacement', gameSessionPlacement_placementId - A unique identifier for a game session placement.

$sel:placedPlayerSessions:GameSessionPlacement', gameSessionPlacement_placedPlayerSessions - A collection of information on player sessions created in response to the game session placement request. These player sessions are created only once a new game session is successfully placed (placement status is FULFILLED). This information includes the player ID (as provided in the placement request) and the corresponding player session ID. Retrieve full player sessions by calling DescribePlayerSessions with the player session ID.

$sel:port:GameSessionPlacement', gameSessionPlacement_port - The port number for the game session. To connect to a GameLift game server, an app needs both the IP address and port number. This value is set once the new game session is placed (placement status is FULFILLED).

$sel:dnsName:GameSessionPlacement', gameSessionPlacement_dnsName - The DNS identifier assigned to the instance that is running the game session. Values have the following format:

TLS-enabled fleets: <unique identifier>.<region identifier>.amazongamelift.com.

Non-TLS-enabled fleets: ec2-<unique identifier>.compute.amazonaws.com. (See Amazon EC2 Instance IP Addressing.)

When connecting to a game session that is running on a TLS-enabled fleet, you must use the DNS name, not the IP address.

$sel:gameSessionRegion:GameSessionPlacement', gameSessionPlacement_gameSessionRegion - Name of the Region where the game session created by this placement request is running. This value is set once the new game session is placed (placement status is FULFILLED).

$sel:playerLatencies:GameSessionPlacement', gameSessionPlacement_playerLatencies - A set of values, expressed in milliseconds, that indicates the amount of latency that a player experiences when connected to AWS Regions.

gameSessionPlacement_gameProperties :: Lens' GameSessionPlacement (Maybe [GameProperty])

A set of custom properties for a game session, formatted as key:value pairs. These properties are passed to a game server process in the GameSession object with a request to start a new game session (see Start a Game Session).

gameSessionPlacement_status :: Lens' GameSessionPlacement (Maybe GameSessionPlacementState)

Current status of the game session placement request.

PENDING -- The placement request is currently in the queue waiting to be processed.

FULFILLED -- A new game session and player sessions (if requested) have been successfully created. Values for GameSessionArn and GameSessionRegion are available.

CANCELLED -- The placement request was canceled with a call to StopGameSessionPlacement.

TIMED_OUT -- A new game session was not successfully created before the time limit expired. You can resubmit the placement request as needed.

FAILED -- GameLift is not able to complete the process of placing the game session. Common reasons are the game session terminated before the placement process was completed, or an unexpected internal error.

gameSessionPlacement_gameSessionQueueName :: Lens' GameSessionPlacement (Maybe Text)

A descriptive label that is associated with game session queue. Queue names must be unique within each Region.

gameSessionPlacement_maximumPlayerSessionCount :: Lens' GameSessionPlacement (Maybe Natural)

The maximum number of players that can be connected simultaneously to the game session.

gameSessionPlacement_matchmakerData :: Lens' GameSessionPlacement (Maybe Text)

Information on the matchmaking process for this game. Data is in JSON syntax, formatted as a string. It identifies the matchmaking configuration used to create the match, and contains data on all players assigned to the match, including player attributes and team assignments. For more details on matchmaker data, see Match Data.

gameSessionPlacement_gameSessionData :: Lens' GameSessionPlacement (Maybe Text)

A set of custom game session properties, formatted as a single string value. This data is passed to a game server process in the GameSession object with a request to start a new game session (see Start a Game Session).

gameSessionPlacement_gameSessionId :: Lens' GameSessionPlacement (Maybe Text)

A unique identifier for the game session. This value is set once the new game session is placed (placement status is FULFILLED).

gameSessionPlacement_startTime :: Lens' GameSessionPlacement (Maybe UTCTime)

Time stamp indicating when this request was placed in the queue. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

gameSessionPlacement_gameSessionArn :: Lens' GameSessionPlacement (Maybe Text)

Identifier for the game session created by this placement request. This value is set once the new game session is placed (placement status is FULFILLED). This identifier is unique across all Regions. You can use this value as a GameSessionId value as needed.

gameSessionPlacement_endTime :: Lens' GameSessionPlacement (Maybe UTCTime)

Time stamp indicating when this request was completed, canceled, or timed out.

gameSessionPlacement_gameSessionName :: Lens' GameSessionPlacement (Maybe Text)

A descriptive label that is associated with a game session. Session names do not need to be unique.

gameSessionPlacement_ipAddress :: Lens' GameSessionPlacement (Maybe Text)

The IP address of the game session. To connect to a GameLift game server, an app needs both the IP address and port number. This value is set once the new game session is placed (placement status is FULFILLED).

gameSessionPlacement_placementId :: Lens' GameSessionPlacement (Maybe Text)

A unique identifier for a game session placement.

gameSessionPlacement_placedPlayerSessions :: Lens' GameSessionPlacement (Maybe [PlacedPlayerSession])

A collection of information on player sessions created in response to the game session placement request. These player sessions are created only once a new game session is successfully placed (placement status is FULFILLED). This information includes the player ID (as provided in the placement request) and the corresponding player session ID. Retrieve full player sessions by calling DescribePlayerSessions with the player session ID.

gameSessionPlacement_port :: Lens' GameSessionPlacement (Maybe Natural)

The port number for the game session. To connect to a GameLift game server, an app needs both the IP address and port number. This value is set once the new game session is placed (placement status is FULFILLED).

gameSessionPlacement_dnsName :: Lens' GameSessionPlacement (Maybe Text)

The DNS identifier assigned to the instance that is running the game session. Values have the following format:

TLS-enabled fleets: <unique identifier>.<region identifier>.amazongamelift.com.

Non-TLS-enabled fleets: ec2-<unique identifier>.compute.amazonaws.com. (See Amazon EC2 Instance IP Addressing.)

When connecting to a game session that is running on a TLS-enabled fleet, you must use the DNS name, not the IP address.

gameSessionPlacement_gameSessionRegion :: Lens' GameSessionPlacement (Maybe Text)

Name of the Region where the game session created by this placement request is running. This value is set once the new game session is placed (placement status is FULFILLED).

gameSessionPlacement_playerLatencies :: Lens' GameSessionPlacement (Maybe [PlayerLatency])

A set of values, expressed in milliseconds, that indicates the amount of latency that a player experiences when connected to AWS Regions.