Network.AWS.GameLift.Types.GameSessionQueue

Description

Documentation

data GameSessionQueue

Configuration for a game session placement mechanism that processes requests for new game sessions. A queue can be used on its own or as part of a matchmaking solution.

Related actions

CreateGameSessionQueue | DescribeGameSessionQueues | UpdateGameSessionQueue

See: newGameSessionQueue smart constructor.

Constructors

GameSessionQueue'

Fields

customEventData :: Maybe Text

Information that is added to all events that are related to this game session queue.

playerLatencyPolicies :: Maybe [PlayerLatencyPolicy]

A set of policies that act as a sliding cap on player latency. FleetIQ works to deliver low latency for most players in a game session. These policies ensure that no individual player can be placed into a game with unreasonably high latency. Use multiple policies to gradually relax latency requirements a step at a time. Multiple policies are applied based on their maximum allowed latency, starting with the lowest value.

priorityConfiguration :: Maybe PriorityConfiguration

Custom settings to use when prioritizing destinations and locations for game session placements. This configuration replaces the FleetIQ default prioritization process. Priority types that are not explicitly named will be automatically applied at the end of the prioritization process.

name :: Maybe Text

A descriptive label that is associated with game session queue. Queue names must be unique within each Region.

destinations :: Maybe [GameSessionQueueDestination]

A list of fleets and/or fleet aliases that can be used to fulfill game session placement requests in the queue. Destinations are identified by either a fleet ARN or a fleet alias ARN, and are listed in order of placement preference.

timeoutInSeconds :: Maybe Natural

The maximum time, in seconds, that a new game session placement request remains in the queue. When a request exceeds this time, the game session placement changes to a TIMED_OUT status.

gameSessionQueueArn :: Maybe Text

The Amazon Resource Name (ARN) that is assigned to a GameLift game session queue resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:<region>::gamesessionqueue/<queue name>. In a GameLift game session queue ARN, the resource ID matches the Name value.

notificationTarget :: Maybe Text

An SNS topic ARN that is set up to receive game session placement notifications. See Setting up notifications for game session placement.

filterConfiguration :: Maybe FilterConfiguration

A list of locations where a queue is allowed to place new game sessions. Locations are specified in the form of AWS Region codes, such as us-west-2. If this parameter is not set, game sessions can be placed in any queue location.

newGameSessionQueue :: GameSessionQueue

Create a value of GameSessionQueue with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:customEventData:GameSessionQueue', gameSessionQueue_customEventData - Information that is added to all events that are related to this game session queue.

$sel:playerLatencyPolicies:GameSessionQueue', gameSessionQueue_playerLatencyPolicies - A set of policies that act as a sliding cap on player latency. FleetIQ works to deliver low latency for most players in a game session. These policies ensure that no individual player can be placed into a game with unreasonably high latency. Use multiple policies to gradually relax latency requirements a step at a time. Multiple policies are applied based on their maximum allowed latency, starting with the lowest value.

$sel:priorityConfiguration:GameSessionQueue', gameSessionQueue_priorityConfiguration - Custom settings to use when prioritizing destinations and locations for game session placements. This configuration replaces the FleetIQ default prioritization process. Priority types that are not explicitly named will be automatically applied at the end of the prioritization process.

$sel:name:GameSessionQueue', gameSessionQueue_name - A descriptive label that is associated with game session queue. Queue names must be unique within each Region.

$sel:destinations:GameSessionQueue', gameSessionQueue_destinations - A list of fleets and/or fleet aliases that can be used to fulfill game session placement requests in the queue. Destinations are identified by either a fleet ARN or a fleet alias ARN, and are listed in order of placement preference.

$sel:timeoutInSeconds:GameSessionQueue', gameSessionQueue_timeoutInSeconds - The maximum time, in seconds, that a new game session placement request remains in the queue. When a request exceeds this time, the game session placement changes to a TIMED_OUT status.

$sel:gameSessionQueueArn:GameSessionQueue', gameSessionQueue_gameSessionQueueArn - The Amazon Resource Name (ARN) that is assigned to a GameLift game session queue resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:<region>::gamesessionqueue/<queue name>. In a GameLift game session queue ARN, the resource ID matches the Name value.

$sel:notificationTarget:GameSessionQueue', gameSessionQueue_notificationTarget - An SNS topic ARN that is set up to receive game session placement notifications. See Setting up notifications for game session placement.

$sel:filterConfiguration:GameSessionQueue', gameSessionQueue_filterConfiguration - A list of locations where a queue is allowed to place new game sessions. Locations are specified in the form of AWS Region codes, such as us-west-2. If this parameter is not set, game sessions can be placed in any queue location.

gameSessionQueue_customEventData :: Lens' GameSessionQueue (Maybe Text)

Information that is added to all events that are related to this game session queue.

gameSessionQueue_playerLatencyPolicies :: Lens' GameSessionQueue (Maybe [PlayerLatencyPolicy])

A set of policies that act as a sliding cap on player latency. FleetIQ works to deliver low latency for most players in a game session. These policies ensure that no individual player can be placed into a game with unreasonably high latency. Use multiple policies to gradually relax latency requirements a step at a time. Multiple policies are applied based on their maximum allowed latency, starting with the lowest value.

gameSessionQueue_priorityConfiguration :: Lens' GameSessionQueue (Maybe PriorityConfiguration)

Custom settings to use when prioritizing destinations and locations for game session placements. This configuration replaces the FleetIQ default prioritization process. Priority types that are not explicitly named will be automatically applied at the end of the prioritization process.

gameSessionQueue_name :: Lens' GameSessionQueue (Maybe Text)

A descriptive label that is associated with game session queue. Queue names must be unique within each Region.

gameSessionQueue_destinations :: Lens' GameSessionQueue (Maybe [GameSessionQueueDestination])

A list of fleets and/or fleet aliases that can be used to fulfill game session placement requests in the queue. Destinations are identified by either a fleet ARN or a fleet alias ARN, and are listed in order of placement preference.

gameSessionQueue_timeoutInSeconds :: Lens' GameSessionQueue (Maybe Natural)

The maximum time, in seconds, that a new game session placement request remains in the queue. When a request exceeds this time, the game session placement changes to a TIMED_OUT status.

gameSessionQueue_gameSessionQueueArn :: Lens' GameSessionQueue (Maybe Text)

The Amazon Resource Name (ARN) that is assigned to a GameLift game session queue resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:<region>::gamesessionqueue/<queue name>. In a GameLift game session queue ARN, the resource ID matches the Name value.

gameSessionQueue_notificationTarget :: Lens' GameSessionQueue (Maybe Text)

An SNS topic ARN that is set up to receive game session placement notifications. See Setting up notifications for game session placement.

gameSessionQueue_filterConfiguration :: Lens' GameSessionQueue (Maybe FilterConfiguration)

A list of locations where a queue is allowed to place new game sessions. Locations are specified in the form of AWS Region codes, such as us-west-2. If this parameter is not set, game sessions can be placed in any queue location.