Network.AWS.GameLift.Types.Instance

Description

Documentation

data Instance

Represents an EC2 instance of virtual computing resources that hosts one or more game servers. In GameLift, a fleet can contain zero or more instances.

Related actions

DescribeInstances

See: newInstance smart constructor.

Constructors

Instance'

Fields

creationTime :: Maybe POSIX

A time stamp indicating when this data object was created. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

instanceId :: Maybe Text

A unique identifier for the instance.

status :: Maybe InstanceStatus

Current status of the instance. Possible statuses include the following:

PENDING -- The instance is in the process of being created and launching server processes as defined in the fleet's run-time configuration.

ACTIVE -- The instance has been successfully created and at least one server process has successfully launched and reported back to GameLift that it is ready to host a game session. The instance is now considered ready to host game sessions.

TERMINATING -- The instance is in the process of shutting down. This may happen to reduce capacity during a scaling down event or to recycle resources in the event of a problem.

ipAddress :: Maybe Text

IP address that is assigned to the instance.

location :: Maybe Text

The fleet location of the instance, expressed as an AWS Region code, such as us-west-2.

operatingSystem :: Maybe OperatingSystem

Operating system that is running on this instance.

fleetArn :: Maybe Text

The Amazon Resource Name (ARN) that is assigned to a GameLift fleet resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:<region>::fleet/fleet-a1234567-b8c9-0d1e-2fa3-b45c6d7e8912.

type' :: Maybe EC2InstanceType

EC2 instance type that defines the computing resources of this instance.

fleetId :: Maybe Text

A unique identifier for the fleet that the instance is in.

dnsName :: Maybe Text

The DNS identifier assigned to the instance that is running the game session. Values have the following format:

TLS-enabled fleets: <unique identifier>.<region identifier>.amazongamelift.com.

Non-TLS-enabled fleets: ec2-<unique identifier>.compute.amazonaws.com. (See Amazon EC2 Instance IP Addressing.)

When connecting to a game session that is running on a TLS-enabled fleet, you must use the DNS name, not the IP address.

newInstance :: Instance

Create a value of Instance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:Instance', instance_creationTime - A time stamp indicating when this data object was created. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

$sel:instanceId:Instance', instance_instanceId - A unique identifier for the instance.

$sel:status:Instance', instance_status - Current status of the instance. Possible statuses include the following:

PENDING -- The instance is in the process of being created and launching server processes as defined in the fleet's run-time configuration.

ACTIVE -- The instance has been successfully created and at least one server process has successfully launched and reported back to GameLift that it is ready to host a game session. The instance is now considered ready to host game sessions.

TERMINATING -- The instance is in the process of shutting down. This may happen to reduce capacity during a scaling down event or to recycle resources in the event of a problem.

$sel:ipAddress:Instance', instance_ipAddress - IP address that is assigned to the instance.

$sel:location:Instance', instance_location - The fleet location of the instance, expressed as an AWS Region code, such as us-west-2.

$sel:operatingSystem:Instance', instance_operatingSystem - Operating system that is running on this instance.

$sel:fleetArn:Instance', instance_fleetArn - The Amazon Resource Name (ARN) that is assigned to a GameLift fleet resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:<region>::fleet/fleet-a1234567-b8c9-0d1e-2fa3-b45c6d7e8912.

$sel:type':Instance', instance_type - EC2 instance type that defines the computing resources of this instance.

$sel:fleetId:Instance', instance_fleetId - A unique identifier for the fleet that the instance is in.

$sel:dnsName:Instance', instance_dnsName - The DNS identifier assigned to the instance that is running the game session. Values have the following format:

TLS-enabled fleets: <unique identifier>.<region identifier>.amazongamelift.com.

Non-TLS-enabled fleets: ec2-<unique identifier>.compute.amazonaws.com. (See Amazon EC2 Instance IP Addressing.)

When connecting to a game session that is running on a TLS-enabled fleet, you must use the DNS name, not the IP address.

instance_creationTime :: Lens' Instance (Maybe UTCTime)

A time stamp indicating when this data object was created. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

instance_instanceId :: Lens' Instance (Maybe Text)

A unique identifier for the instance.

instance_status :: Lens' Instance (Maybe InstanceStatus)

Current status of the instance. Possible statuses include the following:

PENDING -- The instance is in the process of being created and launching server processes as defined in the fleet's run-time configuration.

ACTIVE -- The instance has been successfully created and at least one server process has successfully launched and reported back to GameLift that it is ready to host a game session. The instance is now considered ready to host game sessions.

TERMINATING -- The instance is in the process of shutting down. This may happen to reduce capacity during a scaling down event or to recycle resources in the event of a problem.

instance_ipAddress :: Lens' Instance (Maybe Text)

IP address that is assigned to the instance.

instance_location :: Lens' Instance (Maybe Text)

The fleet location of the instance, expressed as an AWS Region code, such as us-west-2.

instance_operatingSystem :: Lens' Instance (Maybe OperatingSystem)

Operating system that is running on this instance.

instance_fleetArn :: Lens' Instance (Maybe Text)

The Amazon Resource Name (ARN) that is assigned to a GameLift fleet resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:<region>::fleet/fleet-a1234567-b8c9-0d1e-2fa3-b45c6d7e8912.

instance_type :: Lens' Instance (Maybe EC2InstanceType)

EC2 instance type that defines the computing resources of this instance.

instance_fleetId :: Lens' Instance (Maybe Text)

A unique identifier for the fleet that the instance is in.

instance_dnsName :: Lens' Instance (Maybe Text)

The DNS identifier assigned to the instance that is running the game session. Values have the following format:

TLS-enabled fleets: <unique identifier>.<region identifier>.amazongamelift.com.

Non-TLS-enabled fleets: ec2-<unique identifier>.compute.amazonaws.com. (See Amazon EC2 Instance IP Addressing.)

When connecting to a game session that is running on a TLS-enabled fleet, you must use the DNS name, not the IP address.