Network.AWS.GameLift.Types.InstanceAccess

Description

Documentation

data InstanceAccess

Information required to remotely connect to a fleet instance. Access is requested by calling GetInstanceAccess.

See: newInstanceAccess smart constructor.

Constructors

InstanceAccess'

Fields

instanceId :: Maybe Text

A unique identifier for the instance being accessed.

fleetId :: Maybe Text

A unique identifier for the fleet containing the instance being accessed.

ipAddress :: Maybe Text

IP address that is assigned to the instance.

operatingSystem :: Maybe OperatingSystem

Operating system that is running on the instance.

credentials :: Maybe (Sensitive InstanceCredentials)

Credentials required to access the instance.

newInstanceAccess :: InstanceAccess

Create a value of InstanceAccess with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceId:InstanceAccess', instanceAccess_instanceId - A unique identifier for the instance being accessed.

$sel:fleetId:InstanceAccess', instanceAccess_fleetId - A unique identifier for the fleet containing the instance being accessed.

$sel:ipAddress:InstanceAccess', instanceAccess_ipAddress - IP address that is assigned to the instance.

$sel:operatingSystem:InstanceAccess', instanceAccess_operatingSystem - Operating system that is running on the instance.

$sel:credentials:InstanceAccess', instanceAccess_credentials - Credentials required to access the instance.

instanceAccess_instanceId :: Lens' InstanceAccess (Maybe Text)

A unique identifier for the instance being accessed.

instanceAccess_fleetId :: Lens' InstanceAccess (Maybe Text)

A unique identifier for the fleet containing the instance being accessed.

instanceAccess_ipAddress :: Lens' InstanceAccess (Maybe Text)

IP address that is assigned to the instance.

instanceAccess_operatingSystem :: Lens' InstanceAccess (Maybe OperatingSystem)

Operating system that is running on the instance.

instanceAccess_credentials :: Lens' InstanceAccess (Maybe InstanceCredentials)

Credentials required to access the instance.