Network.AWS.GameLift.Types.IpPermission

Description

Documentation

data IpPermission

A range of IP addresses and port settings that allow inbound traffic to connect to server processes on an instance in a fleet. New game sessions are assigned an IP address/port number combination, which must fall into the fleet's allowed ranges. Fleets with custom game builds must have permissions explicitly set. For Realtime Servers fleets, GameLift automatically opens two port ranges, one for TCP messaging and one for UDP.

Related actions

DescribeFleetPortSettings

See: newIpPermission smart constructor.

Constructors

IpPermission'

Fields

fromPort :: Natural

A starting value for a range of allowed port numbers.

toPort :: Natural

An ending value for a range of allowed port numbers. Port numbers are end-inclusive. This value must be higher than FromPort.

ipRange :: Text

A range of allowed IP addresses. This value must be expressed in CIDR notation. Example: "000.000.000.000/[subnet mask]" or optionally the shortened version "0.0.0.0/[subnet mask]".

protocol :: IpProtocol

The network communication protocol used by the fleet.

newIpPermission

Arguments

:: Natural
$sel:fromPort:IpPermission'

-> Natural
$sel:toPort:IpPermission'

-> Text
$sel:ipRange:IpPermission'

-> IpProtocol
$sel:protocol:IpPermission'

-> IpPermission

Create a value of IpPermission with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fromPort:IpPermission', ipPermission_fromPort - A starting value for a range of allowed port numbers.

$sel:toPort:IpPermission', ipPermission_toPort - An ending value for a range of allowed port numbers. Port numbers are end-inclusive. This value must be higher than FromPort.

$sel:ipRange:IpPermission', ipPermission_ipRange - A range of allowed IP addresses. This value must be expressed in CIDR notation. Example: "000.000.000.000/[subnet mask]" or optionally the shortened version "0.0.0.0/[subnet mask]".

$sel:protocol:IpPermission', ipPermission_protocol - The network communication protocol used by the fleet.

ipPermission_fromPort :: Lens' IpPermission Natural

A starting value for a range of allowed port numbers.

ipPermission_toPort :: Lens' IpPermission Natural

An ending value for a range of allowed port numbers. Port numbers are end-inclusive. This value must be higher than FromPort.

ipPermission_ipRange :: Lens' IpPermission Text

A range of allowed IP addresses. This value must be expressed in CIDR notation. Example: "000.000.000.000/[subnet mask]" or optionally the shortened version "0.0.0.0/[subnet mask]".

ipPermission_protocol :: Lens' IpPermission IpProtocol

The network communication protocol used by the fleet.