Network.AWS.GameLift.Types.LocationAttributes

Description

Documentation

data LocationAttributes

Represents a location in a multi-location fleet.

Related actions

DescribeFleetLocationAttributes

See: newLocationAttributes smart constructor.

Constructors

LocationAttributes'

Fields

updateStatus :: Maybe LocationUpdateStatus

The status of fleet activity updates to the location. The status PENDING_UPDATE indicates that StopFleetActions or StartFleetActions has been requested but the update has not yet been completed for the location.

locationState :: Maybe LocationState

A fleet location and its current life-cycle state.

stoppedActions :: Maybe (NonEmpty FleetAction)

A list of fleet actions that have been suspended in the fleet location.

newLocationAttributes :: LocationAttributes

Create a value of LocationAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:updateStatus:LocationAttributes', locationAttributes_updateStatus - The status of fleet activity updates to the location. The status PENDING_UPDATE indicates that StopFleetActions or StartFleetActions has been requested but the update has not yet been completed for the location.

$sel:locationState:LocationAttributes', locationAttributes_locationState - A fleet location and its current life-cycle state.

$sel:stoppedActions:LocationAttributes', locationAttributes_stoppedActions - A list of fleet actions that have been suspended in the fleet location.

locationAttributes_updateStatus :: Lens' LocationAttributes (Maybe LocationUpdateStatus)

The status of fleet activity updates to the location. The status PENDING_UPDATE indicates that StopFleetActions or StartFleetActions has been requested but the update has not yet been completed for the location.

locationAttributes_locationState :: Lens' LocationAttributes (Maybe LocationState)

A fleet location and its current life-cycle state.

locationAttributes_stoppedActions :: Lens' LocationAttributes (Maybe (NonEmpty FleetAction))

A list of fleet actions that have been suspended in the fleet location.