Network.AWS.GameLift.Types.MatchmakingConfiguration

Description

Documentation

data MatchmakingConfiguration

Guidelines for use with FlexMatch to match players into games. All matchmaking requests must specify a matchmaking configuration.

See: newMatchmakingConfiguration smart constructor.

Constructors

MatchmakingConfiguration'

Fields

creationTime :: Maybe POSIX

A time stamp indicating when this data object was created. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

backfillMode :: Maybe BackfillMode

The method used to backfill game sessions created with this matchmaking configuration. MANUAL indicates that the game makes backfill requests or does not use the match backfill feature. AUTOMATIC indicates that GameLift creates StartMatchBackfill requests whenever a game session has one or more open slots. Learn more about manual and automatic backfill in Backfill existing games with FlexMatch. Automatic backfill is not available when FlexMatchMode is set to STANDALONE.

gameProperties :: Maybe [GameProperty]

A set of custom properties for a game session, formatted as key:value pairs. These properties are passed to a game server process in the GameSession object with a request to start a new game session (see Start a Game Session). This information is added to the new GameSession object that is created for a successful match. This parameter is not used when FlexMatchMode is set to STANDALONE.

ruleSetName :: Maybe Text

A unique identifier for the matchmaking rule set to use with this configuration. A matchmaking configuration can only use rule sets that are defined in the same Region.

acceptanceTimeoutSeconds :: Maybe Natural

The length of time (in seconds) to wait for players to accept a proposed match, if acceptance is required. If any player rejects the match or fails to accept before the timeout, the ticket continues to look for an acceptable match.

requestTimeoutSeconds :: Maybe Natural

The maximum duration, in seconds, that a matchmaking ticket can remain in process before timing out. Requests that fail due to timing out can be resubmitted as needed.

notificationTarget :: Maybe Text

An SNS topic ARN that is set up to receive matchmaking notifications.

flexMatchMode :: Maybe FlexMatchMode

Indicates whether this matchmaking configuration is being used with GameLift hosting or as a standalone matchmaking solution.

STANDALONE - FlexMatch forms matches and returns match information, including players and team assignments, in a MatchmakingSucceeded event.

WITH_QUEUE - FlexMatch forms matches and uses the specified GameLift queue to start a game session for the match.

gameSessionQueueArns :: Maybe [Text]

The Amazon Resource Name (ARN) that is assigned to a GameLift game session queue resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:<region>::gamesessionqueue/<queue name>. Queues can be located in any Region. Queues are used to start new GameLift-hosted game sessions for matches that are created with this matchmaking configuration. This property is not set when FlexMatchMode is set to STANDALONE.

name :: Maybe Text

A unique identifier for the matchmaking configuration. This name is used to identify the configuration associated with a matchmaking request or ticket.

customEventData :: Maybe Text

Information to attach to all events related to the matchmaking configuration.

configurationArn :: Maybe Text

The Amazon Resource Name (ARN) that is assigned to a GameLift matchmaking configuration resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:<region>::matchmakingconfiguration/<matchmaking configuration name>. In a GameLift configuration ARN, the resource ID matches the Name value.

acceptanceRequired :: Maybe Bool

A flag that indicates whether a match that was created with this configuration must be accepted by the matched players. To require acceptance, set to TRUE. When this option is enabled, matchmaking tickets use the status REQUIRES_ACCEPTANCE to indicate when a completed potential match is waiting for player acceptance.

gameSessionData :: Maybe Text

A set of custom game session properties, formatted as a single string value. This data is passed to a game server process in the GameSession object with a request to start a new game session (see Start a Game Session). This information is added to the new GameSession object that is created for a successful match. This parameter is not used when FlexMatchMode is set to STANDALONE.

description :: Maybe Text

A descriptive label that is associated with matchmaking configuration.

additionalPlayerCount :: Maybe Natural

The number of player slots in a match to keep open for future players. For example, if the configuration's rule set specifies a match for a single 12-person team, and the additional player count is set to 2, only 10 players are selected for the match. This parameter is not used when FlexMatchMode is set to STANDALONE.

ruleSetArn :: Maybe Text

The Amazon Resource Name (ARN) associated with the GameLift matchmaking rule set resource that this configuration uses.

newMatchmakingConfiguration :: MatchmakingConfiguration

Create a value of MatchmakingConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:MatchmakingConfiguration', matchmakingConfiguration_creationTime - A time stamp indicating when this data object was created. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

$sel:backfillMode:MatchmakingConfiguration', matchmakingConfiguration_backfillMode - The method used to backfill game sessions created with this matchmaking configuration. MANUAL indicates that the game makes backfill requests or does not use the match backfill feature. AUTOMATIC indicates that GameLift creates StartMatchBackfill requests whenever a game session has one or more open slots. Learn more about manual and automatic backfill in Backfill existing games with FlexMatch. Automatic backfill is not available when FlexMatchMode is set to STANDALONE.

$sel:gameProperties:MatchmakingConfiguration', matchmakingConfiguration_gameProperties - A set of custom properties for a game session, formatted as key:value pairs. These properties are passed to a game server process in the GameSession object with a request to start a new game session (see Start a Game Session). This information is added to the new GameSession object that is created for a successful match. This parameter is not used when FlexMatchMode is set to STANDALONE.

$sel:ruleSetName:MatchmakingConfiguration', matchmakingConfiguration_ruleSetName - A unique identifier for the matchmaking rule set to use with this configuration. A matchmaking configuration can only use rule sets that are defined in the same Region.

$sel:acceptanceTimeoutSeconds:MatchmakingConfiguration', matchmakingConfiguration_acceptanceTimeoutSeconds - The length of time (in seconds) to wait for players to accept a proposed match, if acceptance is required. If any player rejects the match or fails to accept before the timeout, the ticket continues to look for an acceptable match.

$sel:requestTimeoutSeconds:MatchmakingConfiguration', matchmakingConfiguration_requestTimeoutSeconds - The maximum duration, in seconds, that a matchmaking ticket can remain in process before timing out. Requests that fail due to timing out can be resubmitted as needed.

$sel:notificationTarget:MatchmakingConfiguration', matchmakingConfiguration_notificationTarget - An SNS topic ARN that is set up to receive matchmaking notifications.

$sel:flexMatchMode:MatchmakingConfiguration', matchmakingConfiguration_flexMatchMode - Indicates whether this matchmaking configuration is being used with GameLift hosting or as a standalone matchmaking solution.

STANDALONE - FlexMatch forms matches and returns match information, including players and team assignments, in a MatchmakingSucceeded event.

WITH_QUEUE - FlexMatch forms matches and uses the specified GameLift queue to start a game session for the match.

$sel:gameSessionQueueArns:MatchmakingConfiguration', matchmakingConfiguration_gameSessionQueueArns - The Amazon Resource Name (ARN) that is assigned to a GameLift game session queue resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:<region>::gamesessionqueue/<queue name>. Queues can be located in any Region. Queues are used to start new GameLift-hosted game sessions for matches that are created with this matchmaking configuration. This property is not set when FlexMatchMode is set to STANDALONE.

$sel:name:MatchmakingConfiguration', matchmakingConfiguration_name - A unique identifier for the matchmaking configuration. This name is used to identify the configuration associated with a matchmaking request or ticket.

$sel:customEventData:MatchmakingConfiguration', matchmakingConfiguration_customEventData - Information to attach to all events related to the matchmaking configuration.

$sel:configurationArn:MatchmakingConfiguration', matchmakingConfiguration_configurationArn - The Amazon Resource Name (ARN) that is assigned to a GameLift matchmaking configuration resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:<region>::matchmakingconfiguration/<matchmaking configuration name>. In a GameLift configuration ARN, the resource ID matches the Name value.

$sel:acceptanceRequired:MatchmakingConfiguration', matchmakingConfiguration_acceptanceRequired - A flag that indicates whether a match that was created with this configuration must be accepted by the matched players. To require acceptance, set to TRUE. When this option is enabled, matchmaking tickets use the status REQUIRES_ACCEPTANCE to indicate when a completed potential match is waiting for player acceptance.

$sel:gameSessionData:MatchmakingConfiguration', matchmakingConfiguration_gameSessionData - A set of custom game session properties, formatted as a single string value. This data is passed to a game server process in the GameSession object with a request to start a new game session (see Start a Game Session). This information is added to the new GameSession object that is created for a successful match. This parameter is not used when FlexMatchMode is set to STANDALONE.

$sel:description:MatchmakingConfiguration', matchmakingConfiguration_description - A descriptive label that is associated with matchmaking configuration.

$sel:additionalPlayerCount:MatchmakingConfiguration', matchmakingConfiguration_additionalPlayerCount - The number of player slots in a match to keep open for future players. For example, if the configuration's rule set specifies a match for a single 12-person team, and the additional player count is set to 2, only 10 players are selected for the match. This parameter is not used when FlexMatchMode is set to STANDALONE.

$sel:ruleSetArn:MatchmakingConfiguration', matchmakingConfiguration_ruleSetArn - The Amazon Resource Name (ARN) associated with the GameLift matchmaking rule set resource that this configuration uses.

matchmakingConfiguration_creationTime :: Lens' MatchmakingConfiguration (Maybe UTCTime)

A time stamp indicating when this data object was created. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

matchmakingConfiguration_backfillMode :: Lens' MatchmakingConfiguration (Maybe BackfillMode)

The method used to backfill game sessions created with this matchmaking configuration. MANUAL indicates that the game makes backfill requests or does not use the match backfill feature. AUTOMATIC indicates that GameLift creates StartMatchBackfill requests whenever a game session has one or more open slots. Learn more about manual and automatic backfill in Backfill existing games with FlexMatch. Automatic backfill is not available when FlexMatchMode is set to STANDALONE.

matchmakingConfiguration_gameProperties :: Lens' MatchmakingConfiguration (Maybe [GameProperty])

A set of custom properties for a game session, formatted as key:value pairs. These properties are passed to a game server process in the GameSession object with a request to start a new game session (see Start a Game Session). This information is added to the new GameSession object that is created for a successful match. This parameter is not used when FlexMatchMode is set to STANDALONE.

matchmakingConfiguration_ruleSetName :: Lens' MatchmakingConfiguration (Maybe Text)

A unique identifier for the matchmaking rule set to use with this configuration. A matchmaking configuration can only use rule sets that are defined in the same Region.

matchmakingConfiguration_acceptanceTimeoutSeconds :: Lens' MatchmakingConfiguration (Maybe Natural)

The length of time (in seconds) to wait for players to accept a proposed match, if acceptance is required. If any player rejects the match or fails to accept before the timeout, the ticket continues to look for an acceptable match.

matchmakingConfiguration_requestTimeoutSeconds :: Lens' MatchmakingConfiguration (Maybe Natural)

The maximum duration, in seconds, that a matchmaking ticket can remain in process before timing out. Requests that fail due to timing out can be resubmitted as needed.

matchmakingConfiguration_notificationTarget :: Lens' MatchmakingConfiguration (Maybe Text)

An SNS topic ARN that is set up to receive matchmaking notifications.

matchmakingConfiguration_flexMatchMode :: Lens' MatchmakingConfiguration (Maybe FlexMatchMode)

Indicates whether this matchmaking configuration is being used with GameLift hosting or as a standalone matchmaking solution.

STANDALONE - FlexMatch forms matches and returns match information, including players and team assignments, in a MatchmakingSucceeded event.

WITH_QUEUE - FlexMatch forms matches and uses the specified GameLift queue to start a game session for the match.

matchmakingConfiguration_gameSessionQueueArns :: Lens' MatchmakingConfiguration (Maybe [Text])

The Amazon Resource Name (ARN) that is assigned to a GameLift game session queue resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:<region>::gamesessionqueue/<queue name>. Queues can be located in any Region. Queues are used to start new GameLift-hosted game sessions for matches that are created with this matchmaking configuration. This property is not set when FlexMatchMode is set to STANDALONE.

matchmakingConfiguration_name :: Lens' MatchmakingConfiguration (Maybe Text)

A unique identifier for the matchmaking configuration. This name is used to identify the configuration associated with a matchmaking request or ticket.

matchmakingConfiguration_customEventData :: Lens' MatchmakingConfiguration (Maybe Text)

Information to attach to all events related to the matchmaking configuration.

matchmakingConfiguration_configurationArn :: Lens' MatchmakingConfiguration (Maybe Text)

The Amazon Resource Name (ARN) that is assigned to a GameLift matchmaking configuration resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:<region>::matchmakingconfiguration/<matchmaking configuration name>. In a GameLift configuration ARN, the resource ID matches the Name value.

matchmakingConfiguration_acceptanceRequired :: Lens' MatchmakingConfiguration (Maybe Bool)

A flag that indicates whether a match that was created with this configuration must be accepted by the matched players. To require acceptance, set to TRUE. When this option is enabled, matchmaking tickets use the status REQUIRES_ACCEPTANCE to indicate when a completed potential match is waiting for player acceptance.

matchmakingConfiguration_gameSessionData :: Lens' MatchmakingConfiguration (Maybe Text)

A set of custom game session properties, formatted as a single string value. This data is passed to a game server process in the GameSession object with a request to start a new game session (see Start a Game Session). This information is added to the new GameSession object that is created for a successful match. This parameter is not used when FlexMatchMode is set to STANDALONE.

matchmakingConfiguration_description :: Lens' MatchmakingConfiguration (Maybe Text)

A descriptive label that is associated with matchmaking configuration.

matchmakingConfiguration_additionalPlayerCount :: Lens' MatchmakingConfiguration (Maybe Natural)

The number of player slots in a match to keep open for future players. For example, if the configuration's rule set specifies a match for a single 12-person team, and the additional player count is set to 2, only 10 players are selected for the match. This parameter is not used when FlexMatchMode is set to STANDALONE.

matchmakingConfiguration_ruleSetArn :: Lens' MatchmakingConfiguration (Maybe Text)

The Amazon Resource Name (ARN) associated with the GameLift matchmaking rule set resource that this configuration uses.