Network.AWS.GameLift.Types.MatchmakingRuleSet

Description

Documentation

data MatchmakingRuleSet

Set of rule statements, used with FlexMatch, that determine how to build your player matches. Each rule set describes a type of group to be created and defines the parameters for acceptable player matches. Rule sets are used in MatchmakingConfiguration objects.

A rule set may define the following elements for a match. For detailed information and examples showing how to construct a rule set, see Build a FlexMatch rule set.

Teams -- Required. A rule set must define one or multiple teams for the match and set minimum and maximum team sizes. For example, a rule set might describe a 4x4 match that requires all eight slots to be filled.

Player attributes -- Optional. These attributes specify a set of player characteristics to evaluate when looking for a match. Matchmaking requests that use a rule set with player attributes must provide the corresponding attribute values. For example, an attribute might specify a player's skill or level.

Rules -- Optional. Rules define how to evaluate potential players for a match based on player attributes. A rule might specify minimum requirements for individual players, teams, or entire matches. For example, a rule might require each player to meet a certain skill level, each team to have at least one player in a certain role, or the match to have a minimum average skill level. or may describe an entire group--such as all teams must be evenly matched or have at least one player in a certain role.

Expansions -- Optional. Expansions allow you to relax the rules after a period of time when no acceptable matches are found. This feature lets you balance getting players into games in a reasonable amount of time instead of making them wait indefinitely for the best possible match. For example, you might use an expansion to increase the maximum skill variance between players after 30 seconds.

See: newMatchmakingRuleSet smart constructor.

Constructors

MatchmakingRuleSet'

Fields

creationTime :: Maybe POSIX

A time stamp indicating when this data object was created. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

ruleSetName :: Maybe Text

A unique identifier for the matchmaking rule set

ruleSetArn :: Maybe Text

The Amazon Resource Name (ARN) that is assigned to a GameLift matchmaking rule set resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:<region>::matchmakingruleset/<ruleset name>. In a GameLift rule set ARN, the resource ID matches the RuleSetName value.

ruleSetBody :: Text

A collection of matchmaking rules, formatted as a JSON string. Comments are not allowed in JSON, but most elements support a description field.

newMatchmakingRuleSet

Arguments

:: Text
$sel:ruleSetBody:MatchmakingRuleSet'

-> MatchmakingRuleSet

Create a value of MatchmakingRuleSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:MatchmakingRuleSet', matchmakingRuleSet_creationTime - A time stamp indicating when this data object was created. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

$sel:ruleSetName:MatchmakingRuleSet', matchmakingRuleSet_ruleSetName - A unique identifier for the matchmaking rule set

$sel:ruleSetArn:MatchmakingRuleSet', matchmakingRuleSet_ruleSetArn - The Amazon Resource Name (ARN) that is assigned to a GameLift matchmaking rule set resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:<region>::matchmakingruleset/<ruleset name>. In a GameLift rule set ARN, the resource ID matches the RuleSetName value.

$sel:ruleSetBody:MatchmakingRuleSet', matchmakingRuleSet_ruleSetBody - A collection of matchmaking rules, formatted as a JSON string. Comments are not allowed in JSON, but most elements support a description field.

matchmakingRuleSet_creationTime :: Lens' MatchmakingRuleSet (Maybe UTCTime)

A time stamp indicating when this data object was created. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

matchmakingRuleSet_ruleSetName :: Lens' MatchmakingRuleSet (Maybe Text)

A unique identifier for the matchmaking rule set

matchmakingRuleSet_ruleSetArn :: Lens' MatchmakingRuleSet (Maybe Text)

The Amazon Resource Name (ARN) that is assigned to a GameLift matchmaking rule set resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:<region>::matchmakingruleset/<ruleset name>. In a GameLift rule set ARN, the resource ID matches the RuleSetName value.

matchmakingRuleSet_ruleSetBody :: Lens' MatchmakingRuleSet Text

A collection of matchmaking rules, formatted as a JSON string. Comments are not allowed in JSON, but most elements support a description field.