Network.AWS.GameLift.Types.PlayerSession

Description

Documentation

data PlayerSession

Represents a player session. Player sessions are created either for a specific game session, or as part of a game session placement or matchmaking request. A player session can represents a reserved player slot in a game session (when status is RESERVED) or actual player activity in a game session (when status is ACTIVE). A player session object, including player data, is automatically passed to a game session when the player connects to the game session and is validated. After the game session ends, player sessions information is retained for 30 days and then removed.

Related actions

CreatePlayerSession | CreatePlayerSessions | DescribePlayerSessions | StartGameSessionPlacement | DescribeGameSessionPlacement | All APIs by task

See: newPlayerSession smart constructor.

Constructors

PlayerSession'

Fields

creationTime :: Maybe POSIX

A time stamp indicating when this data object was created. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

status :: Maybe PlayerSessionStatus

Current status of the player session.

Possible player session statuses include the following:

RESERVED -- The player session request has been received, but the player has not yet connected to the server process and/or been validated.

ACTIVE -- The player has been validated by the server process and is currently connected.

COMPLETED -- The player connection has been dropped.

TIMEDOUT -- A player session request was received, but the player did not connect and/or was not validated within the timeout limit (60 seconds).

playerId :: Maybe Text

A unique identifier for a player that is associated with this player session.

fleetId :: Maybe Text

A unique identifier for the fleet that the player's game session is running on.

fleetArn :: Maybe Text

The Amazon Resource Name (ARN) associated with the GameLift fleet that the player's game session is running on.

playerSessionId :: Maybe Text

A unique identifier for a player session.

gameSessionId :: Maybe Text

A unique identifier for the game session that the player session is connected to.

ipAddress :: Maybe Text

The IP address of the game session. To connect to a GameLift game server, an app needs both the IP address and port number.

terminationTime :: Maybe POSIX

A time stamp indicating when this data object was terminated. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

port :: Maybe Natural

Port number for the game session. To connect to a Amazon GameLift server process, an app needs both the IP address and port number.

dnsName :: Maybe Text

The DNS identifier assigned to the instance that is running the game session. Values have the following format:

TLS-enabled fleets: <unique identifier>.<region identifier>.amazongamelift.com.

Non-TLS-enabled fleets: ec2-<unique identifier>.compute.amazonaws.com. (See Amazon EC2 Instance IP Addressing.)

When connecting to a game session that is running on a TLS-enabled fleet, you must use the DNS name, not the IP address.

playerData :: Maybe Text

Developer-defined information related to a player. GameLift does not use this data, so it can be formatted as needed for use in the game.

newPlayerSession :: PlayerSession

Create a value of PlayerSession with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:PlayerSession', playerSession_creationTime - A time stamp indicating when this data object was created. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

$sel:status:PlayerSession', playerSession_status - Current status of the player session.

Possible player session statuses include the following:

RESERVED -- The player session request has been received, but the player has not yet connected to the server process and/or been validated.

ACTIVE -- The player has been validated by the server process and is currently connected.

COMPLETED -- The player connection has been dropped.

TIMEDOUT -- A player session request was received, but the player did not connect and/or was not validated within the timeout limit (60 seconds).

$sel:playerId:PlayerSession', playerSession_playerId - A unique identifier for a player that is associated with this player session.

$sel:fleetId:PlayerSession', playerSession_fleetId - A unique identifier for the fleet that the player's game session is running on.

$sel:fleetArn:PlayerSession', playerSession_fleetArn - The Amazon Resource Name (ARN) associated with the GameLift fleet that the player's game session is running on.

$sel:playerSessionId:PlayerSession', playerSession_playerSessionId - A unique identifier for a player session.

$sel:gameSessionId:PlayerSession', playerSession_gameSessionId - A unique identifier for the game session that the player session is connected to.

$sel:ipAddress:PlayerSession', playerSession_ipAddress - The IP address of the game session. To connect to a GameLift game server, an app needs both the IP address and port number.

$sel:terminationTime:PlayerSession', playerSession_terminationTime - A time stamp indicating when this data object was terminated. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

$sel:port:PlayerSession', playerSession_port - Port number for the game session. To connect to a Amazon GameLift server process, an app needs both the IP address and port number.

$sel:dnsName:PlayerSession', playerSession_dnsName - The DNS identifier assigned to the instance that is running the game session. Values have the following format:

TLS-enabled fleets: <unique identifier>.<region identifier>.amazongamelift.com.

Non-TLS-enabled fleets: ec2-<unique identifier>.compute.amazonaws.com. (See Amazon EC2 Instance IP Addressing.)

When connecting to a game session that is running on a TLS-enabled fleet, you must use the DNS name, not the IP address.

$sel:playerData:PlayerSession', playerSession_playerData - Developer-defined information related to a player. GameLift does not use this data, so it can be formatted as needed for use in the game.

playerSession_creationTime :: Lens' PlayerSession (Maybe UTCTime)

A time stamp indicating when this data object was created. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

playerSession_status :: Lens' PlayerSession (Maybe PlayerSessionStatus)

Current status of the player session.

Possible player session statuses include the following:

RESERVED -- The player session request has been received, but the player has not yet connected to the server process and/or been validated.

ACTIVE -- The player has been validated by the server process and is currently connected.

COMPLETED -- The player connection has been dropped.

TIMEDOUT -- A player session request was received, but the player did not connect and/or was not validated within the timeout limit (60 seconds).

playerSession_playerId :: Lens' PlayerSession (Maybe Text)

A unique identifier for a player that is associated with this player session.

playerSession_fleetId :: Lens' PlayerSession (Maybe Text)

A unique identifier for the fleet that the player's game session is running on.

playerSession_fleetArn :: Lens' PlayerSession (Maybe Text)

The Amazon Resource Name (ARN) associated with the GameLift fleet that the player's game session is running on.

playerSession_playerSessionId :: Lens' PlayerSession (Maybe Text)

A unique identifier for a player session.

playerSession_gameSessionId :: Lens' PlayerSession (Maybe Text)

A unique identifier for the game session that the player session is connected to.

playerSession_ipAddress :: Lens' PlayerSession (Maybe Text)

The IP address of the game session. To connect to a GameLift game server, an app needs both the IP address and port number.

playerSession_terminationTime :: Lens' PlayerSession (Maybe UTCTime)

A time stamp indicating when this data object was terminated. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

playerSession_port :: Lens' PlayerSession (Maybe Natural)

Port number for the game session. To connect to a Amazon GameLift server process, an app needs both the IP address and port number.

playerSession_dnsName :: Lens' PlayerSession (Maybe Text)

The DNS identifier assigned to the instance that is running the game session. Values have the following format:

TLS-enabled fleets: <unique identifier>.<region identifier>.amazongamelift.com.

Non-TLS-enabled fleets: ec2-<unique identifier>.compute.amazonaws.com. (See Amazon EC2 Instance IP Addressing.)

When connecting to a game session that is running on a TLS-enabled fleet, you must use the DNS name, not the IP address.

playerSession_playerData :: Lens' PlayerSession (Maybe Text)

Developer-defined information related to a player. GameLift does not use this data, so it can be formatted as needed for use in the game.