Network.AWS.GameLift.Types.PriorityConfiguration

Description

Documentation

data PriorityConfiguration

Custom prioritization settings for use by a game session queue when placing new game sessions with available game servers. When defined, this configuration replaces the default FleetIQ prioritization process, which is as follows:

If player latency data is included in a game session request, destinations and locations are prioritized first based on lowest average latency (1), then on lowest hosting cost (2), then on destination list order (3), and finally on location (alphabetical) (4). This approach ensures that the queue's top priority is to place game sessions where average player latency is lowest, and--if latency is the same--where the hosting cost is less, etc.

If player latency data is not included, destinations and locations are prioritized first on destination list order (1), and then on location (alphabetical) (2). This approach ensures that the queue's top priority is to place game sessions on the first destination fleet listed. If that fleet has multiple locations, the game session is placed on the first location (when listed alphabetically).

Changing the priority order will affect how game sessions are placed.

Priority configurations are part of a GameSessionQueue.

See: newPriorityConfiguration smart constructor.

Constructors

PriorityConfiguration'

Fields

locationOrder :: Maybe (NonEmpty Text)

The prioritization order to use for fleet locations, when the PriorityOrder property includes LOCATION. Locations are identified by AWS Region codes such as us-west-2. Each location can only be listed once.

priorityOrder :: Maybe (NonEmpty PriorityType)

The recommended sequence to use when prioritizing where to place new game sessions. Each type can only be listed once.

LATENCY -- FleetIQ prioritizes locations where the average player latency (provided in each game session request) is lowest.

COST -- FleetIQ prioritizes destinations with the lowest current hosting costs. Cost is evaluated based on the location, instance type, and fleet type (Spot or On-Demand) for each destination in the queue.

DESTINATION -- FleetIQ prioritizes based on the order that destinations are listed in the queue configuration.

LOCATION -- FleetIQ prioritizes based on the provided order of locations, as defined in LocationOrder.

newPriorityConfiguration :: PriorityConfiguration

Create a value of PriorityConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:locationOrder:PriorityConfiguration', priorityConfiguration_locationOrder - The prioritization order to use for fleet locations, when the PriorityOrder property includes LOCATION. Locations are identified by AWS Region codes such as us-west-2. Each location can only be listed once.

$sel:priorityOrder:PriorityConfiguration', priorityConfiguration_priorityOrder - The recommended sequence to use when prioritizing where to place new game sessions. Each type can only be listed once.

LATENCY -- FleetIQ prioritizes locations where the average player latency (provided in each game session request) is lowest.

COST -- FleetIQ prioritizes destinations with the lowest current hosting costs. Cost is evaluated based on the location, instance type, and fleet type (Spot or On-Demand) for each destination in the queue.

DESTINATION -- FleetIQ prioritizes based on the order that destinations are listed in the queue configuration.

LOCATION -- FleetIQ prioritizes based on the provided order of locations, as defined in LocationOrder.

priorityConfiguration_locationOrder :: Lens' PriorityConfiguration (Maybe (NonEmpty Text))

The prioritization order to use for fleet locations, when the PriorityOrder property includes LOCATION. Locations are identified by AWS Region codes such as us-west-2. Each location can only be listed once.

priorityConfiguration_priorityOrder :: Lens' PriorityConfiguration (Maybe (NonEmpty PriorityType))

The recommended sequence to use when prioritizing where to place new game sessions. Each type can only be listed once.

LATENCY -- FleetIQ prioritizes locations where the average player latency (provided in each game session request) is lowest.

COST -- FleetIQ prioritizes destinations with the lowest current hosting costs. Cost is evaluated based on the location, instance type, and fleet type (Spot or On-Demand) for each destination in the queue.

DESTINATION -- FleetIQ prioritizes based on the order that destinations are listed in the queue configuration.

LOCATION -- FleetIQ prioritizes based on the provided order of locations, as defined in LocationOrder.