Network.AWS.GameLift.Types.Script

Description

Documentation

data Script

Properties describing a Realtime script.

Related actions

CreateScript | ListScripts | DescribeScript | UpdateScript | DeleteScript | All APIs by task

See: newScript smart constructor.

Constructors

Script'

Fields

creationTime :: Maybe POSIX

A time stamp indicating when this data object was created. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

version :: Maybe Text

Version information that is associated with a build or script. Version strings do not need to be unique.

scriptArn :: Maybe Text

The Amazon Resource Name (ARN) that is assigned to a GameLift script resource and uniquely identifies it. ARNs are unique across all Regions. In a GameLift script ARN, the resource ID matches the ScriptId value.

name :: Maybe Text

A descriptive label that is associated with a script. Script names do not need to be unique.

storageLocation :: Maybe S3Location

sizeOnDisk :: Maybe Natural

The file size of the uploaded Realtime script, expressed in bytes. When files are uploaded from an S3 location, this value remains at "0".

scriptId :: Maybe Text

A unique identifier for the Realtime script

newScript :: Script

Create a value of Script with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:Script', script_creationTime - A time stamp indicating when this data object was created. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

$sel:version:Script', script_version - Version information that is associated with a build or script. Version strings do not need to be unique.

$sel:scriptArn:Script', script_scriptArn - The Amazon Resource Name (ARN) that is assigned to a GameLift script resource and uniquely identifies it. ARNs are unique across all Regions. In a GameLift script ARN, the resource ID matches the ScriptId value.

$sel:name:Script', script_name - A descriptive label that is associated with a script. Script names do not need to be unique.

$sel:storageLocation:Script', script_storageLocation - Undocumented member.

$sel:sizeOnDisk:Script', script_sizeOnDisk - The file size of the uploaded Realtime script, expressed in bytes. When files are uploaded from an S3 location, this value remains at "0".

$sel:scriptId:Script', script_scriptId - A unique identifier for the Realtime script

script_creationTime :: Lens' Script (Maybe UTCTime)

A time stamp indicating when this data object was created. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

script_version :: Lens' Script (Maybe Text)

Version information that is associated with a build or script. Version strings do not need to be unique.

script_scriptArn :: Lens' Script (Maybe Text)

The Amazon Resource Name (ARN) that is assigned to a GameLift script resource and uniquely identifies it. ARNs are unique across all Regions. In a GameLift script ARN, the resource ID matches the ScriptId value.

script_name :: Lens' Script (Maybe Text)

A descriptive label that is associated with a script. Script names do not need to be unique.

script_storageLocation :: Lens' Script (Maybe S3Location)

Undocumented member.

script_sizeOnDisk :: Lens' Script (Maybe Natural)

The file size of the uploaded Realtime script, expressed in bytes. When files are uploaded from an S3 location, this value remains at "0".

script_scriptId :: Lens' Script (Maybe Text)

A unique identifier for the Realtime script