Network.AWS.GameLift.Types.ServerProcess

Description

Documentation

data ServerProcess

A set of instructions for launching server processes on each instance in a fleet. Server processes run either an executable in a custom game build or a Realtime Servers script. Server process configurations are part of a fleet's RuntimeConfiguration.

See: newServerProcess smart constructor.

Constructors

ServerProcess'

Fields

parameters :: Maybe Text

An optional list of parameters to pass to the server executable or Realtime script on launch.

launchPath :: Text

The location of a game build executable or the Realtime script file that contains the Init() function. Game builds and Realtime scripts are installed on instances at the root:

Windows (custom game builds only): C:\game. Example: "C:\game\MyGame\server.exe"

Linux: /local/game. Examples: "/local/game/MyGame/server.exe" or "/local/game/MyRealtimeScript.js"

concurrentExecutions :: Natural

The number of server processes using this configuration that run concurrently on each instance.

newServerProcess

Arguments

:: Text
$sel:launchPath:ServerProcess'

-> Natural
$sel:concurrentExecutions:ServerProcess'

-> ServerProcess

Create a value of ServerProcess with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parameters:ServerProcess', serverProcess_parameters - An optional list of parameters to pass to the server executable or Realtime script on launch.

$sel:launchPath:ServerProcess', serverProcess_launchPath - The location of a game build executable or the Realtime script file that contains the Init() function. Game builds and Realtime scripts are installed on instances at the root:

Windows (custom game builds only): C:\game. Example: "C:\game\MyGame\server.exe"

Linux: /local/game. Examples: "/local/game/MyGame/server.exe" or "/local/game/MyRealtimeScript.js"

$sel:concurrentExecutions:ServerProcess', serverProcess_concurrentExecutions - The number of server processes using this configuration that run concurrently on each instance.

serverProcess_parameters :: Lens' ServerProcess (Maybe Text)

An optional list of parameters to pass to the server executable or Realtime script on launch.

serverProcess_launchPath :: Lens' ServerProcess Text

The location of a game build executable or the Realtime script file that contains the Init() function. Game builds and Realtime scripts are installed on instances at the root:

Windows (custom game builds only): C:\game. Example: "C:\game\MyGame\server.exe"

Linux: /local/game. Examples: "/local/game/MyGame/server.exe" or "/local/game/MyRealtimeScript.js"

serverProcess_concurrentExecutions :: Lens' ServerProcess Natural

The number of server processes using this configuration that run concurrently on each instance.