Network.AWS.GameLift.Types.VpcPeeringConnection

Description

Documentation

data VpcPeeringConnection

Represents a peering connection between a VPC on one of your AWS accounts and the VPC for your Amazon GameLift fleets. This record may be for an active peering connection or a pending connection that has not yet been established.

Related actions

CreateVpcPeeringAuthorization | DescribeVpcPeeringAuthorizations | DeleteVpcPeeringAuthorization | CreateVpcPeeringConnection | DescribeVpcPeeringConnections | DeleteVpcPeeringConnection | All APIs by task

See: newVpcPeeringConnection smart constructor.

Constructors

VpcPeeringConnection'

Fields

vpcPeeringConnectionId :: Maybe Text

A unique identifier that is automatically assigned to the connection record. This ID is referenced in VPC peering connection events, and is used when deleting a connection with DeleteVpcPeeringConnection.

status :: Maybe VpcPeeringConnectionStatus

The status information about the connection. Status indicates if a connection is pending, successful, or failed.

peerVpcId :: Maybe Text

A unique identifier for a VPC with resources to be accessed by your GameLift fleet. The VPC must be in the same Region as your fleet. To look up a VPC ID, use the VPC Dashboard in the AWS Management Console. Learn more about VPC peering in VPC Peering with GameLift Fleets.

fleetArn :: Maybe Text

The Amazon Resource Name (ARN) associated with the GameLift fleet resource for this connection.

ipV4CidrBlock :: Maybe Text

CIDR block of IPv4 addresses assigned to the VPC peering connection for the GameLift VPC. The peered VPC also has an IPv4 CIDR block associated with it; these blocks cannot overlap or the peering connection cannot be created.

gameLiftVpcId :: Maybe Text

A unique identifier for the VPC that contains the Amazon GameLift fleet for this connection. This VPC is managed by Amazon GameLift and does not appear in your AWS account.

fleetId :: Maybe Text

A unique identifier for the fleet. This ID determines the ID of the Amazon GameLift VPC for your fleet.

newVpcPeeringConnection :: VpcPeeringConnection

Create a value of VpcPeeringConnection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpcPeeringConnectionId:VpcPeeringConnection', vpcPeeringConnection_vpcPeeringConnectionId - A unique identifier that is automatically assigned to the connection record. This ID is referenced in VPC peering connection events, and is used when deleting a connection with DeleteVpcPeeringConnection.

$sel:status:VpcPeeringConnection', vpcPeeringConnection_status - The status information about the connection. Status indicates if a connection is pending, successful, or failed.

$sel:peerVpcId:VpcPeeringConnection', vpcPeeringConnection_peerVpcId - A unique identifier for a VPC with resources to be accessed by your GameLift fleet. The VPC must be in the same Region as your fleet. To look up a VPC ID, use the VPC Dashboard in the AWS Management Console. Learn more about VPC peering in VPC Peering with GameLift Fleets.

$sel:fleetArn:VpcPeeringConnection', vpcPeeringConnection_fleetArn - The Amazon Resource Name (ARN) associated with the GameLift fleet resource for this connection.

$sel:ipV4CidrBlock:VpcPeeringConnection', vpcPeeringConnection_ipV4CidrBlock - CIDR block of IPv4 addresses assigned to the VPC peering connection for the GameLift VPC. The peered VPC also has an IPv4 CIDR block associated with it; these blocks cannot overlap or the peering connection cannot be created.

$sel:gameLiftVpcId:VpcPeeringConnection', vpcPeeringConnection_gameLiftVpcId - A unique identifier for the VPC that contains the Amazon GameLift fleet for this connection. This VPC is managed by Amazon GameLift and does not appear in your AWS account.

$sel:fleetId:VpcPeeringConnection', vpcPeeringConnection_fleetId - A unique identifier for the fleet. This ID determines the ID of the Amazon GameLift VPC for your fleet.

vpcPeeringConnection_vpcPeeringConnectionId :: Lens' VpcPeeringConnection (Maybe Text)

A unique identifier that is automatically assigned to the connection record. This ID is referenced in VPC peering connection events, and is used when deleting a connection with DeleteVpcPeeringConnection.

vpcPeeringConnection_status :: Lens' VpcPeeringConnection (Maybe VpcPeeringConnectionStatus)

The status information about the connection. Status indicates if a connection is pending, successful, or failed.

vpcPeeringConnection_peerVpcId :: Lens' VpcPeeringConnection (Maybe Text)

A unique identifier for a VPC with resources to be accessed by your GameLift fleet. The VPC must be in the same Region as your fleet. To look up a VPC ID, use the VPC Dashboard in the AWS Management Console. Learn more about VPC peering in VPC Peering with GameLift Fleets.

vpcPeeringConnection_fleetArn :: Lens' VpcPeeringConnection (Maybe Text)

The Amazon Resource Name (ARN) associated with the GameLift fleet resource for this connection.

vpcPeeringConnection_ipV4CidrBlock :: Lens' VpcPeeringConnection (Maybe Text)

CIDR block of IPv4 addresses assigned to the VPC peering connection for the GameLift VPC. The peered VPC also has an IPv4 CIDR block associated with it; these blocks cannot overlap or the peering connection cannot be created.

vpcPeeringConnection_gameLiftVpcId :: Lens' VpcPeeringConnection (Maybe Text)

A unique identifier for the VPC that contains the Amazon GameLift fleet for this connection. This VPC is managed by Amazon GameLift and does not appear in your AWS account.

vpcPeeringConnection_fleetId :: Lens' VpcPeeringConnection (Maybe Text)

A unique identifier for the fleet. This ID determines the ID of the Amazon GameLift VPC for your fleet.