Network.AWS.GameLift.UpdateBuild

Description

Updates metadata in a build resource, including the build name and version. To update the metadata, specify the build ID to update and provide the new values. If successful, a build object containing the updated metadata is returned.

Learn more

Upload a Custom Server Build

Related actions

CreateBuild | ListBuilds | DescribeBuild | UpdateBuild | DeleteBuild | All APIs by task

Creating a Request

data UpdateBuild

Represents the input for a request operation.

See: newUpdateBuild smart constructor.

Constructors

UpdateBuild'

Fields

version :: Maybe Text

Version information that is associated with a build or script. Version strings do not need to be unique.

name :: Maybe Text

A descriptive label that is associated with a build. Build names do not need to be unique.

buildId :: Text

A unique identifier for the build to update. You can use either the build ID or ARN value.

newUpdateBuild

Arguments

:: Text
$sel:buildId:UpdateBuild'

-> UpdateBuild

Create a value of UpdateBuild with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:version:UpdateBuild', updateBuild_version - Version information that is associated with a build or script. Version strings do not need to be unique.

$sel:name:UpdateBuild', updateBuild_name - A descriptive label that is associated with a build. Build names do not need to be unique.

$sel:buildId:UpdateBuild', updateBuild_buildId - A unique identifier for the build to update. You can use either the build ID or ARN value.

Request Lenses

updateBuild_version :: Lens' UpdateBuild (Maybe Text)

Version information that is associated with a build or script. Version strings do not need to be unique.

updateBuild_name :: Lens' UpdateBuild (Maybe Text)

A descriptive label that is associated with a build. Build names do not need to be unique.

updateBuild_buildId :: Lens' UpdateBuild Text

A unique identifier for the build to update. You can use either the build ID or ARN value.

Destructuring the Response

data UpdateBuildResponse

Represents the returned data in response to a request operation.

See: newUpdateBuildResponse smart constructor.

Constructors

UpdateBuildResponse'

Fields

build :: Maybe Build

The updated build resource.

httpStatus :: Int

The response's http status code.

newUpdateBuildResponse

Arguments

:: Int
$sel:httpStatus:UpdateBuildResponse'

-> UpdateBuildResponse

Create a value of UpdateBuildResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:build:UpdateBuildResponse', updateBuildResponse_build - The updated build resource.

$sel:httpStatus:UpdateBuildResponse', updateBuildResponse_httpStatus - The response's http status code.

Response Lenses

updateBuildResponse_build :: Lens' UpdateBuildResponse (Maybe Build)

The updated build resource.

updateBuildResponse_httpStatus :: Lens' UpdateBuildResponse Int

The response's http status code.