Network.AWS.GameLift.UpdateFleetAttributes

Description

Updates a fleet's mutable attributes, including game session protection and resource creation limits.

To update fleet attributes, specify the fleet ID and the property values that you want to change.

If successful, an updated FleetAttributes object is returned.

Learn more

Setting up GameLift fleets

Related actions

CreateFleetLocations | UpdateFleetAttributes | UpdateFleetCapacity | UpdateFleetPortSettings | UpdateRuntimeConfiguration | StopFleetActions | StartFleetActions | PutScalingPolicy | DeleteFleet | DeleteFleetLocations | DeleteScalingPolicy | All APIs by task

Creating a Request

data UpdateFleetAttributes

Represents the input for a request operation.

See: newUpdateFleetAttributes smart constructor.

Constructors

UpdateFleetAttributes'

Fields

newGameSessionProtectionPolicy' :: Maybe ProtectionPolicy

The game session protection policy to apply to all new instances created in this fleet. Instances that already exist are not affected. You can set protection for individual instances using UpdateGameSession.

NoProtection -- The game session can be terminated during a scale-down event.

FullProtection -- If the game session is in an ACTIVE status, it cannot be terminated during a scale-down event.

name :: Maybe Text

A descriptive label that is associated with a fleet. Fleet names do not need to be unique.

metricGroups :: Maybe [Text]

The name of a metric group to add this fleet to. Use a metric group in Amazon CloudWatch to aggregate the metrics from multiple fleets. Provide an existing metric group name, or create a new metric group by providing a new name. A fleet can only be in one metric group at a time.

description :: Maybe Text

A human-readable description of a fleet.

resourceCreationLimitPolicy :: Maybe ResourceCreationLimitPolicy

Policy settings that limit the number of game sessions an individual player can create over a span of time.

fleetId :: Text

A unique identifier for the fleet to update attribute metadata for. You can use either the fleet ID or ARN value.

newUpdateFleetAttributes

Arguments

:: Text
$sel:fleetId:UpdateFleetAttributes'

-> UpdateFleetAttributes

Create a value of UpdateFleetAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:newGameSessionProtectionPolicy':UpdateFleetAttributes', updateFleetAttributes_newGameSessionProtectionPolicy - The game session protection policy to apply to all new instances created in this fleet. Instances that already exist are not affected. You can set protection for individual instances using UpdateGameSession.

NoProtection -- The game session can be terminated during a scale-down event.

FullProtection -- If the game session is in an ACTIVE status, it cannot be terminated during a scale-down event.

$sel:name:UpdateFleetAttributes', updateFleetAttributes_name - A descriptive label that is associated with a fleet. Fleet names do not need to be unique.

$sel:metricGroups:UpdateFleetAttributes', updateFleetAttributes_metricGroups - The name of a metric group to add this fleet to. Use a metric group in Amazon CloudWatch to aggregate the metrics from multiple fleets. Provide an existing metric group name, or create a new metric group by providing a new name. A fleet can only be in one metric group at a time.

$sel:description:UpdateFleetAttributes', updateFleetAttributes_description - A human-readable description of a fleet.

$sel:resourceCreationLimitPolicy:UpdateFleetAttributes', updateFleetAttributes_resourceCreationLimitPolicy - Policy settings that limit the number of game sessions an individual player can create over a span of time.

$sel:fleetId:UpdateFleetAttributes', updateFleetAttributes_fleetId - A unique identifier for the fleet to update attribute metadata for. You can use either the fleet ID or ARN value.

Request Lenses

updateFleetAttributes_newGameSessionProtectionPolicy :: Lens' UpdateFleetAttributes (Maybe ProtectionPolicy)

The game session protection policy to apply to all new instances created in this fleet. Instances that already exist are not affected. You can set protection for individual instances using UpdateGameSession.

NoProtection -- The game session can be terminated during a scale-down event.

FullProtection -- If the game session is in an ACTIVE status, it cannot be terminated during a scale-down event.

updateFleetAttributes_name :: Lens' UpdateFleetAttributes (Maybe Text)

A descriptive label that is associated with a fleet. Fleet names do not need to be unique.

updateFleetAttributes_metricGroups :: Lens' UpdateFleetAttributes (Maybe [Text])

The name of a metric group to add this fleet to. Use a metric group in Amazon CloudWatch to aggregate the metrics from multiple fleets. Provide an existing metric group name, or create a new metric group by providing a new name. A fleet can only be in one metric group at a time.

updateFleetAttributes_description :: Lens' UpdateFleetAttributes (Maybe Text)

A human-readable description of a fleet.

updateFleetAttributes_resourceCreationLimitPolicy :: Lens' UpdateFleetAttributes (Maybe ResourceCreationLimitPolicy)

Policy settings that limit the number of game sessions an individual player can create over a span of time.

updateFleetAttributes_fleetId :: Lens' UpdateFleetAttributes Text

A unique identifier for the fleet to update attribute metadata for. You can use either the fleet ID or ARN value.

Destructuring the Response

data UpdateFleetAttributesResponse

Represents the returned data in response to a request operation.

See: newUpdateFleetAttributesResponse smart constructor.

Constructors

UpdateFleetAttributesResponse'

Fields

fleetId :: Maybe Text

A unique identifier for the fleet that was updated.

httpStatus :: Int

The response's http status code.

newUpdateFleetAttributesResponse

Arguments

:: Int
$sel:httpStatus:UpdateFleetAttributesResponse'

-> UpdateFleetAttributesResponse

Create a value of UpdateFleetAttributesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fleetId:UpdateFleetAttributes', updateFleetAttributesResponse_fleetId - A unique identifier for the fleet that was updated.

$sel:httpStatus:UpdateFleetAttributesResponse', updateFleetAttributesResponse_httpStatus - The response's http status code.

Response Lenses

updateFleetAttributesResponse_fleetId :: Lens' UpdateFleetAttributesResponse (Maybe Text)

A unique identifier for the fleet that was updated.

updateFleetAttributesResponse_httpStatus :: Lens' UpdateFleetAttributesResponse Int

The response's http status code.