Network.AWS.GameLift.UpdateFleetPortSettings

Description

Updates permissions that allow inbound traffic to connect to game sessions that are being hosted on instances in the fleet.

To update settings, specify the fleet ID to be updated and specify the changes to be made. List the permissions you want to add in InboundPermissionAuthorizations, and permissions you want to remove in InboundPermissionRevocations. Permissions to be removed must match existing fleet permissions.

If successful, the fleet ID for the updated fleet is returned. For fleets with remote locations, port setting updates can take time to propagate across all locations. You can check the status of updates in each location by calling DescribeFleetPortSettings with a location name.

Learn more

Setting up GameLift fleets

Related actions

CreateFleetLocations | UpdateFleetAttributes | UpdateFleetCapacity | UpdateFleetPortSettings | UpdateRuntimeConfiguration | StopFleetActions | StartFleetActions | PutScalingPolicy | DeleteFleet | DeleteFleetLocations | DeleteScalingPolicy | All APIs by task

Creating a Request

data UpdateFleetPortSettings

Represents the input for a request operation.

See: newUpdateFleetPortSettings smart constructor.

Constructors

UpdateFleetPortSettings'

Fields

inboundPermissionAuthorizations :: Maybe [IpPermission]

A collection of port settings to be added to the fleet resource.

inboundPermissionRevocations :: Maybe [IpPermission]

A collection of port settings to be removed from the fleet resource.

fleetId :: Text

A unique identifier for the fleet to update port settings for. You can use either the fleet ID or ARN value.

newUpdateFleetPortSettings

Arguments

:: Text
$sel:fleetId:UpdateFleetPortSettings'

-> UpdateFleetPortSettings

Create a value of UpdateFleetPortSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:inboundPermissionAuthorizations:UpdateFleetPortSettings', updateFleetPortSettings_inboundPermissionAuthorizations - A collection of port settings to be added to the fleet resource.

$sel:inboundPermissionRevocations:UpdateFleetPortSettings', updateFleetPortSettings_inboundPermissionRevocations - A collection of port settings to be removed from the fleet resource.

$sel:fleetId:UpdateFleetPortSettings', updateFleetPortSettings_fleetId - A unique identifier for the fleet to update port settings for. You can use either the fleet ID or ARN value.

Request Lenses

updateFleetPortSettings_inboundPermissionAuthorizations :: Lens' UpdateFleetPortSettings (Maybe [IpPermission])

A collection of port settings to be added to the fleet resource.

updateFleetPortSettings_inboundPermissionRevocations :: Lens' UpdateFleetPortSettings (Maybe [IpPermission])

A collection of port settings to be removed from the fleet resource.

updateFleetPortSettings_fleetId :: Lens' UpdateFleetPortSettings Text

A unique identifier for the fleet to update port settings for. You can use either the fleet ID or ARN value.

Destructuring the Response

data UpdateFleetPortSettingsResponse

Represents the returned data in response to a request operation.

See: newUpdateFleetPortSettingsResponse smart constructor.

Constructors

UpdateFleetPortSettingsResponse'

Fields

fleetId :: Maybe Text

A unique identifier for the fleet that was updated.

httpStatus :: Int

The response's http status code.

newUpdateFleetPortSettingsResponse

Arguments

:: Int
$sel:httpStatus:UpdateFleetPortSettingsResponse'

-> UpdateFleetPortSettingsResponse

Create a value of UpdateFleetPortSettingsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fleetId:UpdateFleetPortSettings', updateFleetPortSettingsResponse_fleetId - A unique identifier for the fleet that was updated.

$sel:httpStatus:UpdateFleetPortSettingsResponse', updateFleetPortSettingsResponse_httpStatus - The response's http status code.

Response Lenses

updateFleetPortSettingsResponse_fleetId :: Lens' UpdateFleetPortSettingsResponse (Maybe Text)

A unique identifier for the fleet that was updated.

updateFleetPortSettingsResponse_httpStatus :: Lens' UpdateFleetPortSettingsResponse Int

The response's http status code.