Network.AWS.GameLift.UpdateGameServer

Description

__This operation is used with the GameLift FleetIQ solution and game server groups.__

Updates information about a registered game server to help GameLift FleetIQ to track game server availability. This operation is called by a game server process that is running on an instance in a game server group.

Use this operation to update the following types of game server information. You can make all three types of updates in the same request:

To update the game server's utilization status, identify the game server and game server group and specify the current utilization status. Use this status to identify when game servers are currently hosting games and when they are available to be claimed.

To report health status, identify the game server and game server group and set health check to HEALTHY. If a game server does not report health status for a certain length of time, the game server is no longer considered healthy. As a result, it will be eventually deregistered from the game server group to avoid affecting utilization metrics. The best practice is to report health every 60 seconds.

To change game server metadata, provide updated game server data.

Once a game server is successfully updated, the relevant statuses and timestamps are updated.

Learn more

GameLift FleetIQ Guide

Related actions

RegisterGameServer | ListGameServers | ClaimGameServer | DescribeGameServer | UpdateGameServer | DeregisterGameServer | All APIs by task

Creating a Request

data UpdateGameServer

See: newUpdateGameServer smart constructor.

Constructors

UpdateGameServer'

Fields

healthCheck :: Maybe GameServerHealthCheck

Indicates health status of the game server. A request that includes this parameter updates the game server's LastHealthCheckTime timestamp.

gameServerData :: Maybe Text

A set of custom game server properties, formatted as a single string value. This data is passed to a game client or service when it requests information on game servers using ListGameServers or ClaimGameServer.

utilizationStatus :: Maybe GameServerUtilizationStatus

Indicates whether the game server is available or is currently hosting gameplay.

gameServerGroupName :: Text

A unique identifier for the game server group where the game server is running. Use either the GameServerGroup name or ARN value.

gameServerId :: Text

A custom string that uniquely identifies the game server to update.

newUpdateGameServer

Arguments

:: Text
$sel:gameServerGroupName:UpdateGameServer'

-> Text
$sel:gameServerId:UpdateGameServer'

-> UpdateGameServer

Create a value of UpdateGameServer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:healthCheck:UpdateGameServer', updateGameServer_healthCheck - Indicates health status of the game server. A request that includes this parameter updates the game server's LastHealthCheckTime timestamp.

$sel:gameServerData:UpdateGameServer', updateGameServer_gameServerData - A set of custom game server properties, formatted as a single string value. This data is passed to a game client or service when it requests information on game servers using ListGameServers or ClaimGameServer.

$sel:utilizationStatus:UpdateGameServer', updateGameServer_utilizationStatus - Indicates whether the game server is available or is currently hosting gameplay.

$sel:gameServerGroupName:UpdateGameServer', updateGameServer_gameServerGroupName - A unique identifier for the game server group where the game server is running. Use either the GameServerGroup name or ARN value.

$sel:gameServerId:UpdateGameServer', updateGameServer_gameServerId - A custom string that uniquely identifies the game server to update.

Request Lenses

updateGameServer_healthCheck :: Lens' UpdateGameServer (Maybe GameServerHealthCheck)

Indicates health status of the game server. A request that includes this parameter updates the game server's LastHealthCheckTime timestamp.

updateGameServer_gameServerData :: Lens' UpdateGameServer (Maybe Text)

A set of custom game server properties, formatted as a single string value. This data is passed to a game client or service when it requests information on game servers using ListGameServers or ClaimGameServer.

updateGameServer_utilizationStatus :: Lens' UpdateGameServer (Maybe GameServerUtilizationStatus)

Indicates whether the game server is available or is currently hosting gameplay.

updateGameServer_gameServerGroupName :: Lens' UpdateGameServer Text

A unique identifier for the game server group where the game server is running. Use either the GameServerGroup name or ARN value.

updateGameServer_gameServerId :: Lens' UpdateGameServer Text

A custom string that uniquely identifies the game server to update.

Destructuring the Response

data UpdateGameServerResponse

See: newUpdateGameServerResponse smart constructor.

Constructors

UpdateGameServerResponse'

Fields

gameServer :: Maybe GameServer

Object that describes the newly updated game server.

httpStatus :: Int

The response's http status code.

newUpdateGameServerResponse

Arguments

:: Int
$sel:httpStatus:UpdateGameServerResponse'

-> UpdateGameServerResponse

Create a value of UpdateGameServerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:gameServer:UpdateGameServerResponse', updateGameServerResponse_gameServer - Object that describes the newly updated game server.

$sel:httpStatus:UpdateGameServerResponse', updateGameServerResponse_httpStatus - The response's http status code.

Response Lenses

updateGameServerResponse_gameServer :: Lens' UpdateGameServerResponse (Maybe GameServer)

Object that describes the newly updated game server.

updateGameServerResponse_httpStatus :: Lens' UpdateGameServerResponse Int

The response's http status code.