Network.AWS.GameLift.UpdateScript

Description

Updates Realtime script metadata and content.

To update script metadata, specify the script ID and provide updated name and/or version values.

To update script content, provide an updated zip file by pointing to either a local file or an Amazon S3 bucket location. You can use either method regardless of how the original script was uploaded. Use the Version parameter to track updates to the script.

If the call is successful, the updated metadata is stored in the script record and a revised script is uploaded to the Amazon GameLift service. Once the script is updated and acquired by a fleet instance, the new version is used for all new game sessions.

Learn more

Amazon GameLift Realtime Servers

Related actions

CreateScript | ListScripts | DescribeScript | UpdateScript | DeleteScript | All APIs by task

Creating a Request

data UpdateScript

See: newUpdateScript smart constructor.

Constructors

UpdateScript'

Fields

zipFile :: Maybe Base64

A data object containing your Realtime scripts and dependencies as a zip file. The zip file can have one or multiple files. Maximum size of a zip file is 5 MB.

When using the AWS CLI tool to create a script, this parameter is set to the zip file name. It must be prepended with the string "fileb://" to indicate that the file data is a binary object. For example: --zip-file fileb://myRealtimeScript.zip.

version :: Maybe Text

Version information that is associated with a build or script. Version strings do not need to be unique.

name :: Maybe Text

A descriptive label that is associated with a script. Script names do not need to be unique.

storageLocation :: Maybe S3Location

The location of the Amazon S3 bucket where a zipped file containing your Realtime scripts is stored. The storage location must specify the Amazon S3 bucket name, the zip file name (the "key"), and a role ARN that allows Amazon GameLift to access the Amazon S3 storage location. The S3 bucket must be in the same Region where you want to create a new script. By default, Amazon GameLift uploads the latest version of the zip file; if you have S3 object versioning turned on, you can use the ObjectVersion parameter to specify an earlier version.

scriptId :: Text

A unique identifier for the Realtime script to update. You can use either the script ID or ARN value.

newUpdateScript

Arguments

:: Text
$sel:scriptId:UpdateScript'

-> UpdateScript

Create a value of UpdateScript with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:zipFile:UpdateScript', updateScript_zipFile - A data object containing your Realtime scripts and dependencies as a zip file. The zip file can have one or multiple files. Maximum size of a zip file is 5 MB.

When using the AWS CLI tool to create a script, this parameter is set to the zip file name. It must be prepended with the string "fileb://" to indicate that the file data is a binary object. For example: --zip-file fileb://myRealtimeScript.zip.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:version:UpdateScript', updateScript_version - Version information that is associated with a build or script. Version strings do not need to be unique.

$sel:name:UpdateScript', updateScript_name - A descriptive label that is associated with a script. Script names do not need to be unique.

$sel:storageLocation:UpdateScript', updateScript_storageLocation - The location of the Amazon S3 bucket where a zipped file containing your Realtime scripts is stored. The storage location must specify the Amazon S3 bucket name, the zip file name (the "key"), and a role ARN that allows Amazon GameLift to access the Amazon S3 storage location. The S3 bucket must be in the same Region where you want to create a new script. By default, Amazon GameLift uploads the latest version of the zip file; if you have S3 object versioning turned on, you can use the ObjectVersion parameter to specify an earlier version.

$sel:scriptId:UpdateScript', updateScript_scriptId - A unique identifier for the Realtime script to update. You can use either the script ID or ARN value.

Request Lenses

updateScript_zipFile :: Lens' UpdateScript (Maybe ByteString)

A data object containing your Realtime scripts and dependencies as a zip file. The zip file can have one or multiple files. Maximum size of a zip file is 5 MB.

When using the AWS CLI tool to create a script, this parameter is set to the zip file name. It must be prepended with the string "fileb://" to indicate that the file data is a binary object. For example: --zip-file fileb://myRealtimeScript.zip.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

updateScript_version :: Lens' UpdateScript (Maybe Text)

Version information that is associated with a build or script. Version strings do not need to be unique.

updateScript_name :: Lens' UpdateScript (Maybe Text)

A descriptive label that is associated with a script. Script names do not need to be unique.

updateScript_storageLocation :: Lens' UpdateScript (Maybe S3Location)

The location of the Amazon S3 bucket where a zipped file containing your Realtime scripts is stored. The storage location must specify the Amazon S3 bucket name, the zip file name (the "key"), and a role ARN that allows Amazon GameLift to access the Amazon S3 storage location. The S3 bucket must be in the same Region where you want to create a new script. By default, Amazon GameLift uploads the latest version of the zip file; if you have S3 object versioning turned on, you can use the ObjectVersion parameter to specify an earlier version.

updateScript_scriptId :: Lens' UpdateScript Text

A unique identifier for the Realtime script to update. You can use either the script ID or ARN value.

Destructuring the Response

data UpdateScriptResponse

See: newUpdateScriptResponse smart constructor.

Constructors

UpdateScriptResponse'

Fields

script :: Maybe Script

The newly created script record with a unique script ID. The new script's storage location reflects an Amazon S3 location: (1) If the script was uploaded from an S3 bucket under your account, the storage location reflects the information that was provided in the CreateScript request; (2) If the script file was uploaded from a local zip file, the storage location reflects an S3 location controls by the Amazon GameLift service.

httpStatus :: Int

The response's http status code.

newUpdateScriptResponse

Arguments

:: Int
$sel:httpStatus:UpdateScriptResponse'

-> UpdateScriptResponse

Create a value of UpdateScriptResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:script:UpdateScriptResponse', updateScriptResponse_script - The newly created script record with a unique script ID. The new script's storage location reflects an Amazon S3 location: (1) If the script was uploaded from an S3 bucket under your account, the storage location reflects the information that was provided in the CreateScript request; (2) If the script file was uploaded from a local zip file, the storage location reflects an S3 location controls by the Amazon GameLift service.

$sel:httpStatus:UpdateScriptResponse', updateScriptResponse_httpStatus - The response's http status code.

Response Lenses

updateScriptResponse_script :: Lens' UpdateScriptResponse (Maybe Script)

The newly created script record with a unique script ID. The new script's storage location reflects an Amazon S3 location: (1) If the script was uploaded from an S3 bucket under your account, the storage location reflects the information that was provided in the CreateScript request; (2) If the script file was uploaded from a local zip file, the storage location reflects an S3 location controls by the Amazon GameLift service.

updateScriptResponse_httpStatus :: Lens' UpdateScriptResponse Int

The response's http status code.