Network.AWS.Glacier.AddTagsToVault

Description

This operation adds the specified tags to a vault. Each tag is composed of a key and a value. Each vault can have up to 10 tags. If your request would cause the tag limit for the vault to be exceeded, the operation throws the LimitExceededException error. If a tag already exists on the vault under a specified key, the existing key value will be overwritten. For more information about tags, see Tagging Amazon S3 Glacier Resources.

Creating a Request

data AddTagsToVault

The input values for AddTagsToVault.

See: newAddTagsToVault smart constructor.

Constructors

AddTagsToVault'

Fields

tags :: Maybe (HashMap Text Text)

The tags to add to the vault. Each tag is composed of a key and a value. The value can be an empty string.

accountId :: Text

The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon S3 Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens ('-') in the ID.

vaultName :: Text

The name of the vault.

newAddTagsToVault

Arguments

:: Text
$sel:accountId:AddTagsToVault'

-> Text
$sel:vaultName:AddTagsToVault'

-> AddTagsToVault

Create a value of AddTagsToVault with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:AddTagsToVault', addTagsToVault_tags - The tags to add to the vault. Each tag is composed of a key and a value. The value can be an empty string.

$sel:accountId:AddTagsToVault', addTagsToVault_accountId - The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon S3 Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens ('-') in the ID.

$sel:vaultName:AddTagsToVault', addTagsToVault_vaultName - The name of the vault.

Request Lenses

addTagsToVault_tags :: Lens' AddTagsToVault (Maybe (HashMap Text Text))

The tags to add to the vault. Each tag is composed of a key and a value. The value can be an empty string.

addTagsToVault_accountId :: Lens' AddTagsToVault Text

The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon S3 Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens ('-') in the ID.

addTagsToVault_vaultName :: Lens' AddTagsToVault Text

The name of the vault.

Destructuring the Response

data AddTagsToVaultResponse

See: newAddTagsToVaultResponse smart constructor.

Constructors

AddTagsToVaultResponse'

newAddTagsToVaultResponse :: AddTagsToVaultResponse

Create a value of AddTagsToVaultResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.